Network.AWS.Glacier.CompleteMultipartUpload

Description

You call this operation to inform Amazon S3 Glacier (Glacier) that all the archive parts have been uploaded and that Glacier can now assemble the archive from the uploaded parts. After assembling and saving the archive to the vault, Glacier returns the URI path of the newly created archive resource. Using the URI path, you can then access the archive. After you upload an archive, you should save the archive ID returned to retrieve the archive at a later point. You can also get the vault inventory to obtain a list of archive IDs in a vault. For more information, see InitiateJob.

In the request, you must include the computed SHA256 tree hash of the entire archive you have uploaded. For information about computing a SHA256 tree hash, see Computing Checksums. On the server side, Glacier also constructs the SHA256 tree hash of the assembled archive. If the values match, Glacier saves the archive to the vault; otherwise, it returns an error, and the operation fails. The ListParts operation returns a list of parts uploaded for a specific multipart upload. It includes checksum information for each uploaded part that can be used to debug a bad checksum issue.

Additionally, Glacier also checks for any missing content ranges when assembling the archive, if missing content ranges are found, Glacier returns an error and the operation fails.

Complete Multipart Upload is an idempotent operation. After your first successful complete multipart upload, if you call the operation again within a short period, the operation will succeed and return the same archive ID. This is useful in the event you experience a network issue that causes an aborted connection or receive a 500 server error, in which case you can repeat your Complete Multipart Upload request and get the same archive ID without creating duplicate archives. Note, however, that after the multipart upload completes, you cannot call the List Parts operation and the multipart upload will not appear in List Multipart Uploads response, even if idempotent complete is possible.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identity and Access Management (IAM).

For conceptual information and underlying REST API, see Uploading Large Archives in Parts (Multipart Upload) and Complete Multipart Upload in the Amazon Glacier Developer Guide.

Creating a Request

data CompleteMultipartUpload

Provides options to complete a multipart upload operation. This informs Amazon Glacier that all the archive parts have been uploaded and Amazon S3 Glacier (Glacier) can now assemble the archive from the uploaded parts. After assembling and saving the archive to the vault, Glacier returns the URI path of the newly created archive resource.

See: newCompleteMultipartUpload smart constructor.

Constructors

CompleteMultipartUpload'

Fields

accountId :: Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

vaultName :: Text

The name of the vault.

uploadId :: Text

The upload ID of the multipart upload.

archiveSize :: Text

The total size, in bytes, of the entire archive. This value should be the sum of all the sizes of the individual parts that you uploaded.

checksum :: Text

The SHA256 tree hash of the entire archive. It is the tree hash of SHA256 tree hash of the individual parts. If the value you specify in the request does not match the SHA256 tree hash of the final assembled archive as computed by Amazon S3 Glacier (Glacier), Glacier returns an error and the request fails.

newCompleteMultipartUpload

Arguments

:: Text
$sel:accountId:CompleteMultipartUpload'

-> Text
$sel:vaultName:CompleteMultipartUpload'

-> Text
$sel:uploadId:CompleteMultipartUpload'

-> Text
$sel:archiveSize:CompleteMultipartUpload'

-> Text
$sel:checksum:CompleteMultipartUpload'

-> CompleteMultipartUpload

Create a value of CompleteMultipartUpload with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:CompleteMultipartUpload', completeMultipartUpload_accountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

$sel:vaultName:CompleteMultipartUpload', completeMultipartUpload_vaultName - The name of the vault.

$sel:uploadId:CompleteMultipartUpload', completeMultipartUpload_uploadId - The upload ID of the multipart upload.

$sel:archiveSize:CompleteMultipartUpload', completeMultipartUpload_archiveSize - The total size, in bytes, of the entire archive. This value should be the sum of all the sizes of the individual parts that you uploaded.

$sel:checksum:CompleteMultipartUpload', completeMultipartUpload_checksum - The SHA256 tree hash of the entire archive. It is the tree hash of SHA256 tree hash of the individual parts. If the value you specify in the request does not match the SHA256 tree hash of the final assembled archive as computed by Amazon S3 Glacier (Glacier), Glacier returns an error and the request fails.

Request Lenses

completeMultipartUpload_accountId :: Lens' CompleteMultipartUpload Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

completeMultipartUpload_vaultName :: Lens' CompleteMultipartUpload Text

The name of the vault.

completeMultipartUpload_uploadId :: Lens' CompleteMultipartUpload Text

The upload ID of the multipart upload.

completeMultipartUpload_archiveSize :: Lens' CompleteMultipartUpload Text

The total size, in bytes, of the entire archive. This value should be the sum of all the sizes of the individual parts that you uploaded.

completeMultipartUpload_checksum :: Lens' CompleteMultipartUpload Text

The SHA256 tree hash of the entire archive. It is the tree hash of SHA256 tree hash of the individual parts. If the value you specify in the request does not match the SHA256 tree hash of the final assembled archive as computed by Amazon S3 Glacier (Glacier), Glacier returns an error and the request fails.

Destructuring the Response

data ArchiveCreationOutput

Contains the Amazon S3 Glacier response to your request.

For information about the underlying REST API, see Upload Archive. For conceptual information, see Working with Archives in Amazon S3 Glacier.

See: newArchiveCreationOutput smart constructor.

Constructors

ArchiveCreationOutput'

Fields

archiveId :: Maybe Text

The ID of the archive. This value is also included as part of the location.

checksum :: Maybe Text

The checksum of the archive computed by Amazon S3 Glacier.

location :: Maybe Text

The relative URI path of the newly added archive resource.

newArchiveCreationOutput :: ArchiveCreationOutput

Create a value of ArchiveCreationOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:archiveId:ArchiveCreationOutput', archiveCreationOutput_archiveId - The ID of the archive. This value is also included as part of the location.

$sel:checksum:ArchiveCreationOutput', archiveCreationOutput_checksum - The checksum of the archive computed by Amazon S3 Glacier.

$sel:location:ArchiveCreationOutput', archiveCreationOutput_location - The relative URI path of the newly added archive resource.

Response Lenses

archiveCreationOutput_archiveId :: Lens' ArchiveCreationOutput (Maybe Text)

The ID of the archive. This value is also included as part of the location.

archiveCreationOutput_checksum :: Lens' ArchiveCreationOutput (Maybe Text)

The checksum of the archive computed by Amazon S3 Glacier.

archiveCreationOutput_location :: Lens' ArchiveCreationOutput (Maybe Text)

The relative URI path of the newly added archive resource.