Network.AWS.Glacier.CompleteVaultLock

Description

This operation completes the vault locking process by transitioning the vault lock from the InProgress state to the Locked state, which causes the vault lock policy to become unchangeable. A vault lock is put into the InProgress state by calling InitiateVaultLock. You can obtain the state of the vault lock by calling GetVaultLock. For more information about the vault locking process, Amazon Glacier Vault Lock.

This operation is idempotent. This request is always successful if the vault lock is in the Locked state and the provided lock ID matches the lock ID originally used to lock the vault.

If an invalid lock ID is passed in the request when the vault lock is in the Locked state, the operation returns an AccessDeniedException error. If an invalid lock ID is passed in the request when the vault lock is in the InProgress state, the operation throws an InvalidParameter error.

Creating a Request

data CompleteVaultLock

The input values for CompleteVaultLock.

See: newCompleteVaultLock smart constructor.

Constructors

CompleteVaultLock'

Fields

accountId :: Text

The AccountId value is the AWS account ID. This value must match the AWS account ID associated with the credentials used to sign the request. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you specify your account ID, do not include any hyphens ('-') in the ID.

vaultName :: Text

The name of the vault.

lockId :: Text

The lockId value is the lock ID obtained from a InitiateVaultLock request.

newCompleteVaultLock

Arguments

:: Text
$sel:accountId:CompleteVaultLock'

-> Text
$sel:vaultName:CompleteVaultLock'

-> Text
$sel:lockId:CompleteVaultLock'

-> CompleteVaultLock

Create a value of CompleteVaultLock with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:CompleteVaultLock', completeVaultLock_accountId - The AccountId value is the AWS account ID. This value must match the AWS account ID associated with the credentials used to sign the request. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you specify your account ID, do not include any hyphens ('-') in the ID.

$sel:vaultName:CompleteVaultLock', completeVaultLock_vaultName - The name of the vault.

$sel:lockId:CompleteVaultLock', completeVaultLock_lockId - The lockId value is the lock ID obtained from a InitiateVaultLock request.

Request Lenses

completeVaultLock_accountId :: Lens' CompleteVaultLock Text

The AccountId value is the AWS account ID. This value must match the AWS account ID associated with the credentials used to sign the request. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you specify your account ID, do not include any hyphens ('-') in the ID.

completeVaultLock_vaultName :: Lens' CompleteVaultLock Text

The name of the vault.

completeVaultLock_lockId :: Lens' CompleteVaultLock Text

The lockId value is the lock ID obtained from a InitiateVaultLock request.

Destructuring the Response

data CompleteVaultLockResponse

See: newCompleteVaultLockResponse smart constructor.

Constructors

CompleteVaultLockResponse'

newCompleteVaultLockResponse :: CompleteVaultLockResponse

Create a value of CompleteVaultLockResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.