Network.AWS.Glacier.DeleteArchive

Description

This operation deletes an archive from a vault. Subsequent requests to initiate a retrieval of this archive will fail. Archive retrievals that are in progress for this archive ID may or may not succeed according to the following scenarios:

If the archive retrieval job is actively preparing the data for download when Amazon S3 Glacier receives the delete archive request, the archival retrieval operation might fail.

If the archive retrieval job has successfully prepared the archive for download when Amazon S3 Glacier receives the delete archive request, you will be able to download the output.

This operation is idempotent. Attempting to delete an already-deleted archive does not result in an error.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identity and Access Management (IAM).

For conceptual information and underlying REST API, see Deleting an Archive in Amazon Glacier and Delete Archive in the Amazon Glacier Developer Guide.

Creating a Request

data DeleteArchive

Provides options for deleting an archive from an Amazon S3 Glacier vault.

See: newDeleteArchive smart constructor.

Constructors

DeleteArchive'

Fields

accountId :: Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

vaultName :: Text

The name of the vault.

archiveId :: Text

The ID of the archive to delete.

newDeleteArchive

Arguments

:: Text
$sel:accountId:DeleteArchive'

-> Text
$sel:vaultName:DeleteArchive'

-> Text
$sel:archiveId:DeleteArchive'

-> DeleteArchive

Create a value of DeleteArchive with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:DeleteArchive', deleteArchive_accountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

$sel:vaultName:DeleteArchive', deleteArchive_vaultName - The name of the vault.

$sel:archiveId:DeleteArchive', deleteArchive_archiveId - The ID of the archive to delete.

Request Lenses

deleteArchive_accountId :: Lens' DeleteArchive Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

deleteArchive_vaultName :: Lens' DeleteArchive Text

The name of the vault.

deleteArchive_archiveId :: Lens' DeleteArchive Text

The ID of the archive to delete.

Destructuring the Response

data DeleteArchiveResponse

See: newDeleteArchiveResponse smart constructor.

Constructors

DeleteArchiveResponse'

newDeleteArchiveResponse :: DeleteArchiveResponse

Create a value of DeleteArchiveResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.