Network.AWS.Glacier.DescribeVault

Description

This operation returns information about a vault, including the vault's Amazon Resource Name (ARN), the date the vault was created, the number of archives it contains, and the total size of all the archives in the vault. The number of archives and their total size are as of the last inventory generation. This means that if you add or remove an archive from a vault, and then immediately use Describe Vault, the change in contents will not be immediately reflected. If you want to retrieve the latest inventory of the vault, use InitiateJob. Amazon S3 Glacier generates vault inventories approximately daily. For more information, see Downloading a Vault Inventory in Amazon S3 Glacier.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identity and Access Management (IAM).

For conceptual information and underlying REST API, see Retrieving Vault Metadata in Amazon S3 Glacier and Describe Vault in the Amazon Glacier Developer Guide.

Creating a Request

data DescribeVault

Provides options for retrieving metadata for a specific vault in Amazon Glacier.

See: newDescribeVault smart constructor.

Constructors

DescribeVault'

Fields

accountId :: Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

vaultName :: Text

The name of the vault.

newDescribeVault

Arguments

:: Text
$sel:accountId:DescribeVault'

-> Text
$sel:vaultName:DescribeVault'

-> DescribeVault

Create a value of DescribeVault with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:DescribeVault', describeVault_accountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

$sel:vaultName:DescribeVault', describeVault_vaultName - The name of the vault.

Request Lenses

describeVault_accountId :: Lens' DescribeVault Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

describeVault_vaultName :: Lens' DescribeVault Text

The name of the vault.

Destructuring the Response

data DescribeVaultOutput

Contains the Amazon S3 Glacier response to your request.

See: newDescribeVaultOutput smart constructor.

Constructors

DescribeVaultOutput'

Fields

lastInventoryDate :: Maybe Text

The Universal Coordinated Time (UTC) date when Amazon S3 Glacier completed the last vault inventory. This value should be a string in the ISO 8601 date format, for example 2012-03-20T17:03:43.221Z.

creationDate :: Maybe Text

The Universal Coordinated Time (UTC) date when the vault was created. This value should be a string in the ISO 8601 date format, for example 2012-03-20T17:03:43.221Z.

vaultName :: Maybe Text

The name of the vault.

vaultARN :: Maybe Text

The Amazon Resource Name (ARN) of the vault.

sizeInBytes :: Maybe Integer

Total size, in bytes, of the archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example if you just created the vault.

numberOfArchives :: Maybe Integer

The number of archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example if you just created the vault.

newDescribeVaultOutput :: DescribeVaultOutput

Create a value of DescribeVaultOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastInventoryDate:DescribeVaultOutput', describeVaultOutput_lastInventoryDate - The Universal Coordinated Time (UTC) date when Amazon S3 Glacier completed the last vault inventory. This value should be a string in the ISO 8601 date format, for example 2012-03-20T17:03:43.221Z.

$sel:creationDate:DescribeVaultOutput', describeVaultOutput_creationDate - The Universal Coordinated Time (UTC) date when the vault was created. This value should be a string in the ISO 8601 date format, for example 2012-03-20T17:03:43.221Z.

$sel:vaultName:DescribeVaultOutput', describeVaultOutput_vaultName - The name of the vault.

$sel:vaultARN:DescribeVaultOutput', describeVaultOutput_vaultARN - The Amazon Resource Name (ARN) of the vault.

$sel:sizeInBytes:DescribeVaultOutput', describeVaultOutput_sizeInBytes - Total size, in bytes, of the archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example if you just created the vault.

$sel:numberOfArchives:DescribeVaultOutput', describeVaultOutput_numberOfArchives - The number of archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example if you just created the vault.

Response Lenses

describeVaultOutput_lastInventoryDate :: Lens' DescribeVaultOutput (Maybe Text)

The Universal Coordinated Time (UTC) date when Amazon S3 Glacier completed the last vault inventory. This value should be a string in the ISO 8601 date format, for example 2012-03-20T17:03:43.221Z.

describeVaultOutput_creationDate :: Lens' DescribeVaultOutput (Maybe Text)

The Universal Coordinated Time (UTC) date when the vault was created. This value should be a string in the ISO 8601 date format, for example 2012-03-20T17:03:43.221Z.

describeVaultOutput_vaultName :: Lens' DescribeVaultOutput (Maybe Text)

The name of the vault.

describeVaultOutput_vaultARN :: Lens' DescribeVaultOutput (Maybe Text)

The Amazon Resource Name (ARN) of the vault.

describeVaultOutput_sizeInBytes :: Lens' DescribeVaultOutput (Maybe Integer)

Total size, in bytes, of the archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example if you just created the vault.

describeVaultOutput_numberOfArchives :: Lens' DescribeVaultOutput (Maybe Integer)

The number of archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example if you just created the vault.