Network.AWS.Glacier.GetDataRetrievalPolicy

Description

This operation returns the current data retrieval policy for the account and region specified in the GET request. For more information about data retrieval policies, see Amazon Glacier Data Retrieval Policies.

Creating a Request

data GetDataRetrievalPolicy

Input for GetDataRetrievalPolicy.

See: newGetDataRetrievalPolicy smart constructor.

Constructors

GetDataRetrievalPolicy'

Fields

accountId :: Text

The AccountId value is the AWS account ID. This value must match the AWS account ID associated with the credentials used to sign the request. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you specify your account ID, do not include any hyphens ('-') in the ID.

newGetDataRetrievalPolicy

Arguments

:: Text
$sel:accountId:GetDataRetrievalPolicy'

-> GetDataRetrievalPolicy

Create a value of GetDataRetrievalPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:GetDataRetrievalPolicy', getDataRetrievalPolicy_accountId - The AccountId value is the AWS account ID. This value must match the AWS account ID associated with the credentials used to sign the request. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you specify your account ID, do not include any hyphens ('-') in the ID.

Request Lenses

getDataRetrievalPolicy_accountId :: Lens' GetDataRetrievalPolicy Text

The AccountId value is the AWS account ID. This value must match the AWS account ID associated with the credentials used to sign the request. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon Glacier uses the AWS account ID associated with the credentials used to sign the request. If you specify your account ID, do not include any hyphens ('-') in the ID.

Destructuring the Response

data GetDataRetrievalPolicyResponse

Contains the Amazon S3 Glacier response to the GetDataRetrievalPolicy request.

See: newGetDataRetrievalPolicyResponse smart constructor.

Constructors

GetDataRetrievalPolicyResponse'

Fields

policy :: Maybe DataRetrievalPolicy

Contains the returned data retrieval policy in JSON format.

httpStatus :: Int

The response's http status code.

newGetDataRetrievalPolicyResponse

Arguments

:: Int
$sel:httpStatus:GetDataRetrievalPolicyResponse'

-> GetDataRetrievalPolicyResponse

Create a value of GetDataRetrievalPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policy:GetDataRetrievalPolicyResponse', getDataRetrievalPolicyResponse_policy - Contains the returned data retrieval policy in JSON format.

$sel:httpStatus:GetDataRetrievalPolicyResponse', getDataRetrievalPolicyResponse_httpStatus - The response's http status code.

Response Lenses

getDataRetrievalPolicyResponse_policy :: Lens' GetDataRetrievalPolicyResponse (Maybe DataRetrievalPolicy)

Contains the returned data retrieval policy in JSON format.

getDataRetrievalPolicyResponse_httpStatus :: Lens' GetDataRetrievalPolicyResponse Int

The response's http status code.