Network.AWS.Glacier.GetVaultNotifications

Description

This operation retrieves the notification-configuration subresource of the specified vault.

For information about setting a notification configuration on a vault, see SetVaultNotifications. If a notification configuration for a vault is not set, the operation returns a 404 Not Found error. For more information about vault notifications, see Configuring Vault Notifications in Amazon S3 Glacier.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identity and Access Management (IAM).

For conceptual information and underlying REST API, see Configuring Vault Notifications in Amazon S3 Glacier and Get Vault Notification Configuration in the Amazon Glacier Developer Guide.

Creating a Request

data GetVaultNotifications

Provides options for retrieving the notification configuration set on an Amazon Glacier vault.

See: newGetVaultNotifications smart constructor.

Constructors

GetVaultNotifications'

Fields

accountId :: Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

vaultName :: Text

The name of the vault.

newGetVaultNotifications

Arguments

:: Text
$sel:accountId:GetVaultNotifications'

-> Text
$sel:vaultName:GetVaultNotifications'

-> GetVaultNotifications

Create a value of GetVaultNotifications with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:GetVaultNotifications', getVaultNotifications_accountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

$sel:vaultName:GetVaultNotifications', getVaultNotifications_vaultName - The name of the vault.

Request Lenses

getVaultNotifications_accountId :: Lens' GetVaultNotifications Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

getVaultNotifications_vaultName :: Lens' GetVaultNotifications Text

The name of the vault.

Destructuring the Response

data GetVaultNotificationsResponse

Contains the Amazon S3 Glacier response to your request.

See: newGetVaultNotificationsResponse smart constructor.

Constructors

GetVaultNotificationsResponse'

Fields

vaultNotificationConfig :: Maybe VaultNotificationConfig

Returns the notification configuration set on the vault.

httpStatus :: Int

The response's http status code.

newGetVaultNotificationsResponse

Arguments

:: Int
$sel:httpStatus:GetVaultNotificationsResponse'

-> GetVaultNotificationsResponse

Create a value of GetVaultNotificationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vaultNotificationConfig:GetVaultNotificationsResponse', getVaultNotificationsResponse_vaultNotificationConfig - Returns the notification configuration set on the vault.

$sel:httpStatus:GetVaultNotificationsResponse', getVaultNotificationsResponse_httpStatus - The response's http status code.

Response Lenses

getVaultNotificationsResponse_vaultNotificationConfig :: Lens' GetVaultNotificationsResponse (Maybe VaultNotificationConfig)

Returns the notification configuration set on the vault.

getVaultNotificationsResponse_httpStatus :: Lens' GetVaultNotificationsResponse Int

The response's http status code.