Network.AWS.Glacier.InitiateJob

Description

This operation initiates a job of the specified type, which can be a select, an archival retrieval, or a vault retrieval. For more information about using this operation, see the documentation for the underlying REST API Initiate a Job.

Creating a Request

data InitiateJob

Provides options for initiating an Amazon S3 Glacier job.

See: newInitiateJob smart constructor.

Constructors

InitiateJob'

Fields

jobParameters :: Maybe JobParameters

Provides options for specifying job information.

accountId :: Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

vaultName :: Text

The name of the vault.

newInitiateJob

Arguments

:: Text
$sel:accountId:InitiateJob'

-> Text
$sel:vaultName:InitiateJob'

-> InitiateJob

Create a value of InitiateJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobParameters:InitiateJob', initiateJob_jobParameters - Provides options for specifying job information.

$sel:accountId:InitiateJob', initiateJob_accountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

$sel:vaultName:InitiateJob', initiateJob_vaultName - The name of the vault.

Request Lenses

initiateJob_jobParameters :: Lens' InitiateJob (Maybe JobParameters)

Provides options for specifying job information.

initiateJob_accountId :: Lens' InitiateJob Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

initiateJob_vaultName :: Lens' InitiateJob Text

The name of the vault.

Destructuring the Response

data InitiateJobResponse

Contains the Amazon S3 Glacier response to your request.

See: newInitiateJobResponse smart constructor.

Constructors

InitiateJobResponse'

Fields

jobOutputPath :: Maybe Text

The path to the location of where the select results are stored.

location :: Maybe Text

The relative URI path of the job.

jobId :: Maybe Text

The ID of the job.

httpStatus :: Int

The response's http status code.

newInitiateJobResponse

Arguments

:: Int
$sel:httpStatus:InitiateJobResponse'

-> InitiateJobResponse

Create a value of InitiateJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobOutputPath:InitiateJobResponse', initiateJobResponse_jobOutputPath - The path to the location of where the select results are stored.

$sel:location:InitiateJobResponse', initiateJobResponse_location - The relative URI path of the job.

$sel:jobId:InitiateJobResponse', initiateJobResponse_jobId - The ID of the job.

$sel:httpStatus:InitiateJobResponse', initiateJobResponse_httpStatus - The response's http status code.

Response Lenses

initiateJobResponse_jobOutputPath :: Lens' InitiateJobResponse (Maybe Text)

The path to the location of where the select results are stored.

initiateJobResponse_location :: Lens' InitiateJobResponse (Maybe Text)

The relative URI path of the job.

initiateJobResponse_jobId :: Lens' InitiateJobResponse (Maybe Text)

The ID of the job.

initiateJobResponse_httpStatus :: Lens' InitiateJobResponse Int

The response's http status code.