Network.AWS.Glacier.InitiateMultipartUpload

Description

This operation initiates a multipart upload. Amazon S3 Glacier creates a multipart upload resource and returns its ID in the response. The multipart upload ID is used in subsequent requests to upload parts of an archive (see UploadMultipartPart).

When you initiate a multipart upload, you specify the part size in number of bytes. The part size must be a megabyte (1024 KB) multiplied by a power of 2-for example, 1048576 (1 MB), 2097152 (2 MB), 4194304 (4 MB), 8388608 (8 MB), and so on. The minimum allowable part size is 1 MB, and the maximum is 4 GB.

Every part you upload to this resource (see UploadMultipartPart), except the last one, must have the same size. The last one can be the same size or smaller. For example, suppose you want to upload a 16.2 MB file. If you initiate the multipart upload with a part size of 4 MB, you will upload four parts of 4 MB each and one part of 0.2 MB.

You don't need to know the size of the archive when you start a multipart upload because Amazon S3 Glacier does not require you to specify the overall archive size.

After you complete the multipart upload, Amazon S3 Glacier (Glacier) removes the multipart upload resource referenced by the ID. Glacier also removes the multipart upload resource if you cancel the multipart upload or it may be removed if there is no activity for a period of 24 hours.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identity and Access Management (IAM).

For conceptual information and underlying REST API, see Uploading Large Archives in Parts (Multipart Upload) and Initiate Multipart Upload in the Amazon Glacier Developer Guide.

Creating a Request

data InitiateMultipartUpload

Provides options for initiating a multipart upload to an Amazon S3 Glacier vault.

See: newInitiateMultipartUpload smart constructor.

Constructors

InitiateMultipartUpload'

Fields

archiveDescription :: Maybe Text

The archive description that you are uploading in parts.

The part size must be a megabyte (1024 KB) multiplied by a power of 2, for example 1048576 (1 MB), 2097152 (2 MB), 4194304 (4 MB), 8388608 (8 MB), and so on. The minimum allowable part size is 1 MB, and the maximum is 4 GB (4096 MB).

accountId :: Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

vaultName :: Text

The name of the vault.

partSize :: Text

The size of each part except the last, in bytes. The last part can be smaller than this part size.

newInitiateMultipartUpload

Arguments

:: Text
$sel:accountId:InitiateMultipartUpload'

-> Text
$sel:vaultName:InitiateMultipartUpload'

-> Text
$sel:partSize:InitiateMultipartUpload'

-> InitiateMultipartUpload

Create a value of InitiateMultipartUpload with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:archiveDescription:InitiateMultipartUpload', initiateMultipartUpload_archiveDescription - The archive description that you are uploading in parts.

The part size must be a megabyte (1024 KB) multiplied by a power of 2, for example 1048576 (1 MB), 2097152 (2 MB), 4194304 (4 MB), 8388608 (8 MB), and so on. The minimum allowable part size is 1 MB, and the maximum is 4 GB (4096 MB).

$sel:accountId:InitiateMultipartUpload', initiateMultipartUpload_accountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

$sel:vaultName:InitiateMultipartUpload', initiateMultipartUpload_vaultName - The name of the vault.

$sel:partSize:InitiateMultipartUpload', initiateMultipartUpload_partSize - The size of each part except the last, in bytes. The last part can be smaller than this part size.

Request Lenses

initiateMultipartUpload_archiveDescription :: Lens' InitiateMultipartUpload (Maybe Text)

The archive description that you are uploading in parts.

The part size must be a megabyte (1024 KB) multiplied by a power of 2, for example 1048576 (1 MB), 2097152 (2 MB), 4194304 (4 MB), 8388608 (8 MB), and so on. The minimum allowable part size is 1 MB, and the maximum is 4 GB (4096 MB).

initiateMultipartUpload_accountId :: Lens' InitiateMultipartUpload Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

initiateMultipartUpload_vaultName :: Lens' InitiateMultipartUpload Text

The name of the vault.

initiateMultipartUpload_partSize :: Lens' InitiateMultipartUpload Text

The size of each part except the last, in bytes. The last part can be smaller than this part size.

Destructuring the Response

data InitiateMultipartUploadResponse

The Amazon S3 Glacier response to your request.

See: newInitiateMultipartUploadResponse smart constructor.

Constructors

InitiateMultipartUploadResponse'

Fields

uploadId :: Maybe Text

The ID of the multipart upload. This value is also included as part of the location.

location :: Maybe Text

The relative URI path of the multipart upload ID Amazon S3 Glacier created.

httpStatus :: Int

The response's http status code.

newInitiateMultipartUploadResponse

Arguments

:: Int
$sel:httpStatus:InitiateMultipartUploadResponse'

-> InitiateMultipartUploadResponse

Create a value of InitiateMultipartUploadResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:uploadId:InitiateMultipartUploadResponse', initiateMultipartUploadResponse_uploadId - The ID of the multipart upload. This value is also included as part of the location.

$sel:location:InitiateMultipartUploadResponse', initiateMultipartUploadResponse_location - The relative URI path of the multipart upload ID Amazon S3 Glacier created.

$sel:httpStatus:InitiateMultipartUploadResponse', initiateMultipartUploadResponse_httpStatus - The response's http status code.

Response Lenses

initiateMultipartUploadResponse_uploadId :: Lens' InitiateMultipartUploadResponse (Maybe Text)

The ID of the multipart upload. This value is also included as part of the location.

initiateMultipartUploadResponse_location :: Lens' InitiateMultipartUploadResponse (Maybe Text)

The relative URI path of the multipart upload ID Amazon S3 Glacier created.

initiateMultipartUploadResponse_httpStatus :: Lens' InitiateMultipartUploadResponse Int

The response's http status code.