Network.AWS.Glacier.ListTagsForVault

Description

This operation lists all the tags attached to a vault. The operation returns an empty map if there are no tags. For more information about tags, see Tagging Amazon S3 Glacier Resources.

Creating a Request

data ListTagsForVault

The input value for ListTagsForVaultInput.

See: newListTagsForVault smart constructor.

Constructors

ListTagsForVault'

Fields

accountId :: Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

vaultName :: Text

The name of the vault.

newListTagsForVault

Arguments

:: Text
$sel:accountId:ListTagsForVault'

-> Text
$sel:vaultName:ListTagsForVault'

-> ListTagsForVault

Create a value of ListTagsForVault with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:ListTagsForVault', listTagsForVault_accountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

$sel:vaultName:ListTagsForVault', listTagsForVault_vaultName - The name of the vault.

Request Lenses

listTagsForVault_accountId :: Lens' ListTagsForVault Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

listTagsForVault_vaultName :: Lens' ListTagsForVault Text

The name of the vault.

Destructuring the Response

data ListTagsForVaultResponse

Contains the Amazon S3 Glacier response to your request.

See: newListTagsForVaultResponse smart constructor.

Constructors

ListTagsForVaultResponse'

Fields

tags :: Maybe (HashMap Text Text)

The tags attached to the vault. Each tag is composed of a key and a value.

httpStatus :: Int

The response's http status code.

newListTagsForVaultResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForVaultResponse'

-> ListTagsForVaultResponse

Create a value of ListTagsForVaultResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForVaultResponse', listTagsForVaultResponse_tags - The tags attached to the vault. Each tag is composed of a key and a value.

$sel:httpStatus:ListTagsForVaultResponse', listTagsForVaultResponse_httpStatus - The response's http status code.

Response Lenses

listTagsForVaultResponse_tags :: Lens' ListTagsForVaultResponse (Maybe (HashMap Text Text))

The tags attached to the vault. Each tag is composed of a key and a value.

listTagsForVaultResponse_httpStatus :: Lens' ListTagsForVaultResponse Int

The response's http status code.