Network.AWS.Glacier.Types.CSVInput

Description

Documentation

data CSVInput

Contains information about the comma-separated value (CSV) file to select from.

See: newCSVInput smart constructor.

Constructors

CSVInput'

Fields

quoteCharacter :: Maybe Text

A value used as an escape character where the field delimiter is part of the value.

recordDelimiter :: Maybe Text

A value used to separate individual records from each other.

fileHeaderInfo :: Maybe FileHeaderInfo

Describes the first line of input. Valid values are None, Ignore, and Use.

quoteEscapeCharacter :: Maybe Text

A single character used for escaping the quotation-mark character inside an already escaped value.

comments :: Maybe Text

A single character used to indicate that a row should be ignored when the character is present at the start of that row.

fieldDelimiter :: Maybe Text

A value used to separate individual fields from each other within a record.

newCSVInput :: CSVInput

Create a value of CSVInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:quoteCharacter:CSVInput', cSVInput_quoteCharacter - A value used as an escape character where the field delimiter is part of the value.

$sel:recordDelimiter:CSVInput', cSVInput_recordDelimiter - A value used to separate individual records from each other.

$sel:fileHeaderInfo:CSVInput', cSVInput_fileHeaderInfo - Describes the first line of input. Valid values are None, Ignore, and Use.

$sel:quoteEscapeCharacter:CSVInput', cSVInput_quoteEscapeCharacter - A single character used for escaping the quotation-mark character inside an already escaped value.

$sel:comments:CSVInput', cSVInput_comments - A single character used to indicate that a row should be ignored when the character is present at the start of that row.

$sel:fieldDelimiter:CSVInput', cSVInput_fieldDelimiter - A value used to separate individual fields from each other within a record.

cSVInput_quoteCharacter :: Lens' CSVInput (Maybe Text)

A value used as an escape character where the field delimiter is part of the value.

cSVInput_recordDelimiter :: Lens' CSVInput (Maybe Text)

A value used to separate individual records from each other.

cSVInput_fileHeaderInfo :: Lens' CSVInput (Maybe FileHeaderInfo)

Describes the first line of input. Valid values are None, Ignore, and Use.

cSVInput_quoteEscapeCharacter :: Lens' CSVInput (Maybe Text)

A single character used for escaping the quotation-mark character inside an already escaped value.

cSVInput_comments :: Lens' CSVInput (Maybe Text)

A single character used to indicate that a row should be ignored when the character is present at the start of that row.

cSVInput_fieldDelimiter :: Lens' CSVInput (Maybe Text)

A value used to separate individual fields from each other within a record.