Network.AWS.Glacier.Types.DescribeVaultOutput

Description

Documentation

data DescribeVaultOutput

Contains the Amazon S3 Glacier response to your request.

See: newDescribeVaultOutput smart constructor.

Constructors

DescribeVaultOutput'

Fields

lastInventoryDate :: Maybe Text

The Universal Coordinated Time (UTC) date when Amazon S3 Glacier completed the last vault inventory. This value should be a string in the ISO 8601 date format, for example 2012-03-20T17:03:43.221Z.

creationDate :: Maybe Text

The Universal Coordinated Time (UTC) date when the vault was created. This value should be a string in the ISO 8601 date format, for example 2012-03-20T17:03:43.221Z.

vaultName :: Maybe Text

The name of the vault.

vaultARN :: Maybe Text

The Amazon Resource Name (ARN) of the vault.

sizeInBytes :: Maybe Integer

Total size, in bytes, of the archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example if you just created the vault.

numberOfArchives :: Maybe Integer

The number of archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example if you just created the vault.

newDescribeVaultOutput :: DescribeVaultOutput

Create a value of DescribeVaultOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastInventoryDate:DescribeVaultOutput', describeVaultOutput_lastInventoryDate - The Universal Coordinated Time (UTC) date when Amazon S3 Glacier completed the last vault inventory. This value should be a string in the ISO 8601 date format, for example 2012-03-20T17:03:43.221Z.

$sel:creationDate:DescribeVaultOutput', describeVaultOutput_creationDate - The Universal Coordinated Time (UTC) date when the vault was created. This value should be a string in the ISO 8601 date format, for example 2012-03-20T17:03:43.221Z.

$sel:vaultName:DescribeVaultOutput', describeVaultOutput_vaultName - The name of the vault.

$sel:vaultARN:DescribeVaultOutput', describeVaultOutput_vaultARN - The Amazon Resource Name (ARN) of the vault.

$sel:sizeInBytes:DescribeVaultOutput', describeVaultOutput_sizeInBytes - Total size, in bytes, of the archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example if you just created the vault.

$sel:numberOfArchives:DescribeVaultOutput', describeVaultOutput_numberOfArchives - The number of archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example if you just created the vault.

describeVaultOutput_lastInventoryDate :: Lens' DescribeVaultOutput (Maybe Text)

The Universal Coordinated Time (UTC) date when Amazon S3 Glacier completed the last vault inventory. This value should be a string in the ISO 8601 date format, for example 2012-03-20T17:03:43.221Z.

describeVaultOutput_creationDate :: Lens' DescribeVaultOutput (Maybe Text)

The Universal Coordinated Time (UTC) date when the vault was created. This value should be a string in the ISO 8601 date format, for example 2012-03-20T17:03:43.221Z.

describeVaultOutput_vaultName :: Lens' DescribeVaultOutput (Maybe Text)

The name of the vault.

describeVaultOutput_vaultARN :: Lens' DescribeVaultOutput (Maybe Text)

The Amazon Resource Name (ARN) of the vault.

describeVaultOutput_sizeInBytes :: Lens' DescribeVaultOutput (Maybe Integer)

Total size, in bytes, of the archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example if you just created the vault.

describeVaultOutput_numberOfArchives :: Lens' DescribeVaultOutput (Maybe Integer)

The number of archives in the vault as of the last inventory date. This field will return null if an inventory has not yet run on the vault, for example if you just created the vault.