Network.AWS.Glacier.Types.InventoryRetrievalJobDescription

Description

Documentation

data InventoryRetrievalJobDescription

Describes the options for a range inventory retrieval job.

See: newInventoryRetrievalJobDescription smart constructor.

Constructors

InventoryRetrievalJobDescription'

Fields

startDate :: Maybe Text

The start of the date range in Universal Coordinated Time (UTC) for vault inventory retrieval that includes archives created on or after this date. This value should be a string in the ISO 8601 date format, for example 2013-03-20T17:03:43Z.

format :: Maybe Text

The output format for the vault inventory list, which is set by the InitiateJob request when initiating a job to retrieve a vault inventory. Valid values are CSV and JSON.

endDate :: Maybe Text

The end of the date range in UTC for vault inventory retrieval that includes archives created before this date. This value should be a string in the ISO 8601 date format, for example 2013-03-20T17:03:43Z.

limit :: Maybe Text

The maximum number of inventory items returned per vault inventory retrieval request. This limit is set when initiating the job with the a InitiateJob request.

marker :: Maybe Text

An opaque string that represents where to continue pagination of the vault inventory retrieval results. You use the marker in a new InitiateJob request to obtain additional inventory items. If there are no more inventory items, this value is null. For more information, see Range Inventory Retrieval.

newInventoryRetrievalJobDescription :: InventoryRetrievalJobDescription

Create a value of InventoryRetrievalJobDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startDate:InventoryRetrievalJobDescription', inventoryRetrievalJobDescription_startDate - The start of the date range in Universal Coordinated Time (UTC) for vault inventory retrieval that includes archives created on or after this date. This value should be a string in the ISO 8601 date format, for example 2013-03-20T17:03:43Z.

$sel:format:InventoryRetrievalJobDescription', inventoryRetrievalJobDescription_format - The output format for the vault inventory list, which is set by the InitiateJob request when initiating a job to retrieve a vault inventory. Valid values are CSV and JSON.

$sel:endDate:InventoryRetrievalJobDescription', inventoryRetrievalJobDescription_endDate - The end of the date range in UTC for vault inventory retrieval that includes archives created before this date. This value should be a string in the ISO 8601 date format, for example 2013-03-20T17:03:43Z.

$sel:limit:InventoryRetrievalJobDescription', inventoryRetrievalJobDescription_limit - The maximum number of inventory items returned per vault inventory retrieval request. This limit is set when initiating the job with the a InitiateJob request.

$sel:marker:InventoryRetrievalJobDescription', inventoryRetrievalJobDescription_marker - An opaque string that represents where to continue pagination of the vault inventory retrieval results. You use the marker in a new InitiateJob request to obtain additional inventory items. If there are no more inventory items, this value is null. For more information, see Range Inventory Retrieval.

inventoryRetrievalJobDescription_startDate :: Lens' InventoryRetrievalJobDescription (Maybe Text)

The start of the date range in Universal Coordinated Time (UTC) for vault inventory retrieval that includes archives created on or after this date. This value should be a string in the ISO 8601 date format, for example 2013-03-20T17:03:43Z.

inventoryRetrievalJobDescription_format :: Lens' InventoryRetrievalJobDescription (Maybe Text)

The output format for the vault inventory list, which is set by the InitiateJob request when initiating a job to retrieve a vault inventory. Valid values are CSV and JSON.

inventoryRetrievalJobDescription_endDate :: Lens' InventoryRetrievalJobDescription (Maybe Text)

The end of the date range in UTC for vault inventory retrieval that includes archives created before this date. This value should be a string in the ISO 8601 date format, for example 2013-03-20T17:03:43Z.

inventoryRetrievalJobDescription_limit :: Lens' InventoryRetrievalJobDescription (Maybe Text)

The maximum number of inventory items returned per vault inventory retrieval request. This limit is set when initiating the job with the a InitiateJob request.

inventoryRetrievalJobDescription_marker :: Lens' InventoryRetrievalJobDescription (Maybe Text)

An opaque string that represents where to continue pagination of the vault inventory retrieval results. You use the marker in a new InitiateJob request to obtain additional inventory items. If there are no more inventory items, this value is null. For more information, see Range Inventory Retrieval.