Network.AWS.Glacier.Types.InventoryRetrievalJobInput

Description

Documentation

data InventoryRetrievalJobInput

Provides options for specifying a range inventory retrieval job.

See: newInventoryRetrievalJobInput smart constructor.

Constructors

InventoryRetrievalJobInput'

Fields

startDate :: Maybe Text

The start of the date range in UTC for vault inventory retrieval that includes archives created on or after this date. This value should be a string in the ISO 8601 date format, for example 2013-03-20T17:03:43Z.

endDate :: Maybe Text

The end of the date range in UTC for vault inventory retrieval that includes archives created before this date. This value should be a string in the ISO 8601 date format, for example 2013-03-20T17:03:43Z.

limit :: Maybe Text

Specifies the maximum number of inventory items returned per vault inventory retrieval request. Valid values are greater than or equal to 1.

marker :: Maybe Text

An opaque string that represents where to continue pagination of the vault inventory retrieval results. You use the marker in a new InitiateJob request to obtain additional inventory items. If there are no more inventory items, this value is null.

newInventoryRetrievalJobInput :: InventoryRetrievalJobInput

Create a value of InventoryRetrievalJobInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startDate:InventoryRetrievalJobInput', inventoryRetrievalJobInput_startDate - The start of the date range in UTC for vault inventory retrieval that includes archives created on or after this date. This value should be a string in the ISO 8601 date format, for example 2013-03-20T17:03:43Z.

$sel:endDate:InventoryRetrievalJobInput', inventoryRetrievalJobInput_endDate - The end of the date range in UTC for vault inventory retrieval that includes archives created before this date. This value should be a string in the ISO 8601 date format, for example 2013-03-20T17:03:43Z.

$sel:limit:InventoryRetrievalJobInput', inventoryRetrievalJobInput_limit - Specifies the maximum number of inventory items returned per vault inventory retrieval request. Valid values are greater than or equal to 1.

$sel:marker:InventoryRetrievalJobInput', inventoryRetrievalJobInput_marker - An opaque string that represents where to continue pagination of the vault inventory retrieval results. You use the marker in a new InitiateJob request to obtain additional inventory items. If there are no more inventory items, this value is null.

inventoryRetrievalJobInput_startDate :: Lens' InventoryRetrievalJobInput (Maybe Text)

The start of the date range in UTC for vault inventory retrieval that includes archives created on or after this date. This value should be a string in the ISO 8601 date format, for example 2013-03-20T17:03:43Z.

inventoryRetrievalJobInput_endDate :: Lens' InventoryRetrievalJobInput (Maybe Text)

The end of the date range in UTC for vault inventory retrieval that includes archives created before this date. This value should be a string in the ISO 8601 date format, for example 2013-03-20T17:03:43Z.

inventoryRetrievalJobInput_limit :: Lens' InventoryRetrievalJobInput (Maybe Text)

Specifies the maximum number of inventory items returned per vault inventory retrieval request. Valid values are greater than or equal to 1.

inventoryRetrievalJobInput_marker :: Lens' InventoryRetrievalJobInput (Maybe Text)

An opaque string that represents where to continue pagination of the vault inventory retrieval results. You use the marker in a new InitiateJob request to obtain additional inventory items. If there are no more inventory items, this value is null.