Network.AWS.Glacier.Types.JobParameters

Description

Documentation

data JobParameters

Provides options for defining a job.

See: newJobParameters smart constructor.

Constructors

JobParameters'

Fields

format :: Maybe Text

When initiating a job to retrieve a vault inventory, you can optionally add this parameter to your request to specify the output format. If you are initiating an inventory job and do not specify a Format field, JSON is the default format. Valid values are "CSV" and "JSON".

retrievalByteRange :: Maybe Text

The byte range to retrieve for an archive retrieval. in the form "StartByteValue-EndByteValue" If not specified, the whole archive is retrieved. If specified, the byte range must be megabyte (1024*1024) aligned which means that StartByteValue must be divisible by 1 MB and EndByteValue plus 1 must be divisible by 1 MB or be the end of the archive specified as the archive byte size value minus 1. If RetrievalByteRange is not megabyte aligned, this operation returns a 400 response.

An error occurs if you specify this field for an inventory retrieval job request.

selectParameters :: Maybe SelectParameters

Contains the parameters that define a job.

sNSTopic :: Maybe Text

The Amazon SNS topic ARN to which Amazon S3 Glacier sends a notification when the job is completed and the output is ready for you to download. The specified topic publishes the notification to its subscribers. The SNS topic must exist.

archiveId :: Maybe Text

The ID of the archive that you want to retrieve. This field is required only if Type is set to select or archive-retrievalcode>. An error occurs if you specify this request parameter for an inventory retrieval job request.

description :: Maybe Text

The optional description for the job. The description must be less than or equal to 1,024 bytes. The allowable characters are 7-bit ASCII without control codes-specifically, ASCII values 32-126 decimal or 0x20-0x7E hexadecimal.

inventoryRetrievalParameters :: Maybe InventoryRetrievalJobInput

Input parameters used for range inventory retrieval.

type' :: Maybe Text

The job type. You can initiate a job to perform a select query on an archive, retrieve an archive, or get an inventory of a vault. Valid values are "select", "archive-retrieval" and "inventory-retrieval".

tier :: Maybe Text

The tier to use for a select or an archive retrieval job. Valid values are Expedited, Standard, or Bulk. Standard is the default.

outputLocation :: Maybe OutputLocation

Contains information about the location where the select job results are stored.

newJobParameters :: JobParameters

Create a value of JobParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:format:JobParameters', jobParameters_format - When initiating a job to retrieve a vault inventory, you can optionally add this parameter to your request to specify the output format. If you are initiating an inventory job and do not specify a Format field, JSON is the default format. Valid values are "CSV" and "JSON".

$sel:retrievalByteRange:JobParameters', jobParameters_retrievalByteRange - The byte range to retrieve for an archive retrieval. in the form "StartByteValue-EndByteValue" If not specified, the whole archive is retrieved. If specified, the byte range must be megabyte (1024*1024) aligned which means that StartByteValue must be divisible by 1 MB and EndByteValue plus 1 must be divisible by 1 MB or be the end of the archive specified as the archive byte size value minus 1. If RetrievalByteRange is not megabyte aligned, this operation returns a 400 response.

An error occurs if you specify this field for an inventory retrieval job request.

$sel:selectParameters:JobParameters', jobParameters_selectParameters - Contains the parameters that define a job.

$sel:sNSTopic:JobParameters', jobParameters_sNSTopic - The Amazon SNS topic ARN to which Amazon S3 Glacier sends a notification when the job is completed and the output is ready for you to download. The specified topic publishes the notification to its subscribers. The SNS topic must exist.

$sel:archiveId:JobParameters', jobParameters_archiveId - The ID of the archive that you want to retrieve. This field is required only if Type is set to select or archive-retrievalcode>. An error occurs if you specify this request parameter for an inventory retrieval job request.

$sel:description:JobParameters', jobParameters_description - The optional description for the job. The description must be less than or equal to 1,024 bytes. The allowable characters are 7-bit ASCII without control codes-specifically, ASCII values 32-126 decimal or 0x20-0x7E hexadecimal.

$sel:inventoryRetrievalParameters:JobParameters', jobParameters_inventoryRetrievalParameters - Input parameters used for range inventory retrieval.

$sel:type':JobParameters', jobParameters_type - The job type. You can initiate a job to perform a select query on an archive, retrieve an archive, or get an inventory of a vault. Valid values are "select", "archive-retrieval" and "inventory-retrieval".

$sel:tier:JobParameters', jobParameters_tier - The tier to use for a select or an archive retrieval job. Valid values are Expedited, Standard, or Bulk. Standard is the default.

$sel:outputLocation:JobParameters', jobParameters_outputLocation - Contains information about the location where the select job results are stored.

jobParameters_format :: Lens' JobParameters (Maybe Text)

When initiating a job to retrieve a vault inventory, you can optionally add this parameter to your request to specify the output format. If you are initiating an inventory job and do not specify a Format field, JSON is the default format. Valid values are "CSV" and "JSON".

jobParameters_retrievalByteRange :: Lens' JobParameters (Maybe Text)

The byte range to retrieve for an archive retrieval. in the form "StartByteValue-EndByteValue" If not specified, the whole archive is retrieved. If specified, the byte range must be megabyte (1024*1024) aligned which means that StartByteValue must be divisible by 1 MB and EndByteValue plus 1 must be divisible by 1 MB or be the end of the archive specified as the archive byte size value minus 1. If RetrievalByteRange is not megabyte aligned, this operation returns a 400 response.

An error occurs if you specify this field for an inventory retrieval job request.

jobParameters_selectParameters :: Lens' JobParameters (Maybe SelectParameters)

Contains the parameters that define a job.

jobParameters_sNSTopic :: Lens' JobParameters (Maybe Text)

The Amazon SNS topic ARN to which Amazon S3 Glacier sends a notification when the job is completed and the output is ready for you to download. The specified topic publishes the notification to its subscribers. The SNS topic must exist.

jobParameters_archiveId :: Lens' JobParameters (Maybe Text)

The ID of the archive that you want to retrieve. This field is required only if Type is set to select or archive-retrievalcode>. An error occurs if you specify this request parameter for an inventory retrieval job request.

jobParameters_description :: Lens' JobParameters (Maybe Text)

The optional description for the job. The description must be less than or equal to 1,024 bytes. The allowable characters are 7-bit ASCII without control codes-specifically, ASCII values 32-126 decimal or 0x20-0x7E hexadecimal.

jobParameters_inventoryRetrievalParameters :: Lens' JobParameters (Maybe InventoryRetrievalJobInput)

Input parameters used for range inventory retrieval.

jobParameters_type :: Lens' JobParameters (Maybe Text)

The job type. You can initiate a job to perform a select query on an archive, retrieve an archive, or get an inventory of a vault. Valid values are "select", "archive-retrieval" and "inventory-retrieval".

jobParameters_tier :: Lens' JobParameters (Maybe Text)

The tier to use for a select or an archive retrieval job. Valid values are Expedited, Standard, or Bulk. Standard is the default.

jobParameters_outputLocation :: Lens' JobParameters (Maybe OutputLocation)

Contains information about the location where the select job results are stored.