Network.AWS.Glacier.Types.PartListElement

Description

Documentation

data PartListElement

A list of the part sizes of the multipart upload.

See: newPartListElement smart constructor.

Constructors

PartListElement'

Fields

sHA256TreeHash :: Maybe Text

The SHA256 tree hash value that Amazon S3 Glacier calculated for the part. This field is never null.

rangeInBytes :: Maybe Text

The byte range of a part, inclusive of the upper value of the range.

newPartListElement :: PartListElement

Create a value of PartListElement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sHA256TreeHash:PartListElement', partListElement_sHA256TreeHash - The SHA256 tree hash value that Amazon S3 Glacier calculated for the part. This field is never null.

$sel:rangeInBytes:PartListElement', partListElement_rangeInBytes - The byte range of a part, inclusive of the upper value of the range.

partListElement_sHA256TreeHash :: Lens' PartListElement (Maybe Text)

The SHA256 tree hash value that Amazon S3 Glacier calculated for the part. This field is never null.

partListElement_rangeInBytes :: Lens' PartListElement (Maybe Text)

The byte range of a part, inclusive of the upper value of the range.