Network.AWS.Glacier.Types.ProvisionedCapacityDescription

Description

Documentation

data ProvisionedCapacityDescription

The definition for a provisioned capacity unit.

See: newProvisionedCapacityDescription smart constructor.

Constructors

ProvisionedCapacityDescription'

Fields

capacityId :: Maybe Text

The ID that identifies the provisioned capacity unit.

startDate :: Maybe Text

The date that the provisioned capacity unit was purchased, in Universal Coordinated Time (UTC).

expirationDate :: Maybe Text

The date that the provisioned capacity unit expires, in Universal Coordinated Time (UTC).

newProvisionedCapacityDescription :: ProvisionedCapacityDescription

Create a value of ProvisionedCapacityDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:capacityId:ProvisionedCapacityDescription', provisionedCapacityDescription_capacityId - The ID that identifies the provisioned capacity unit.

$sel:startDate:ProvisionedCapacityDescription', provisionedCapacityDescription_startDate - The date that the provisioned capacity unit was purchased, in Universal Coordinated Time (UTC).

$sel:expirationDate:ProvisionedCapacityDescription', provisionedCapacityDescription_expirationDate - The date that the provisioned capacity unit expires, in Universal Coordinated Time (UTC).

provisionedCapacityDescription_capacityId :: Lens' ProvisionedCapacityDescription (Maybe Text)

The ID that identifies the provisioned capacity unit.

provisionedCapacityDescription_startDate :: Lens' ProvisionedCapacityDescription (Maybe Text)

The date that the provisioned capacity unit was purchased, in Universal Coordinated Time (UTC).

provisionedCapacityDescription_expirationDate :: Lens' ProvisionedCapacityDescription (Maybe Text)

The date that the provisioned capacity unit expires, in Universal Coordinated Time (UTC).