Network.AWS.Glacier.Types.S3Location

Description

Documentation

data S3Location

Contains information about the location in Amazon S3 where the select job results are stored.

See: newS3Location smart constructor.

Constructors

S3Location'

Fields

cannedACL :: Maybe CannedACL

The canned access control list (ACL) to apply to the job results.

prefix :: Maybe Text

The prefix that is prepended to the results for this request.

bucketName :: Maybe Text

The name of the Amazon S3 bucket where the job results are stored.

accessControlList :: Maybe [Grant]

A list of grants that control access to the staged results.

userMetadata :: Maybe (HashMap Text Text)

A map of metadata to store with the job results in Amazon S3.

encryption :: Maybe Encryption

Contains information about the encryption used to store the job results in Amazon S3.

storageClass :: Maybe StorageClass

The storage class used to store the job results.

tagging :: Maybe (HashMap Text Text)

The tag-set that is applied to the job results.

newS3Location :: S3Location

Create a value of S3Location with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cannedACL:S3Location', s3Location_cannedACL - The canned access control list (ACL) to apply to the job results.

$sel:prefix:S3Location', s3Location_prefix - The prefix that is prepended to the results for this request.

$sel:bucketName:S3Location', s3Location_bucketName - The name of the Amazon S3 bucket where the job results are stored.

$sel:accessControlList:S3Location', s3Location_accessControlList - A list of grants that control access to the staged results.

$sel:userMetadata:S3Location', s3Location_userMetadata - A map of metadata to store with the job results in Amazon S3.

$sel:encryption:S3Location', s3Location_encryption - Contains information about the encryption used to store the job results in Amazon S3.

$sel:storageClass:S3Location', s3Location_storageClass - The storage class used to store the job results.

$sel:tagging:S3Location', s3Location_tagging - The tag-set that is applied to the job results.

s3Location_cannedACL :: Lens' S3Location (Maybe CannedACL)

The canned access control list (ACL) to apply to the job results.

s3Location_prefix :: Lens' S3Location (Maybe Text)

The prefix that is prepended to the results for this request.

s3Location_bucketName :: Lens' S3Location (Maybe Text)

The name of the Amazon S3 bucket where the job results are stored.

s3Location_accessControlList :: Lens' S3Location (Maybe [Grant])

A list of grants that control access to the staged results.

s3Location_userMetadata :: Lens' S3Location (Maybe (HashMap Text Text))

A map of metadata to store with the job results in Amazon S3.

s3Location_encryption :: Lens' S3Location (Maybe Encryption)

Contains information about the encryption used to store the job results in Amazon S3.

s3Location_storageClass :: Lens' S3Location (Maybe StorageClass)

The storage class used to store the job results.

s3Location_tagging :: Lens' S3Location (Maybe (HashMap Text Text))

The tag-set that is applied to the job results.