Network.AWS.Glacier.Types.UploadListElement

Description

Documentation

data UploadListElement

A list of in-progress multipart uploads for a vault.

See: newUploadListElement smart constructor.

Constructors

UploadListElement'

Fields

partSizeInBytes :: Maybe Integer

The part size, in bytes, specified in the Initiate Multipart Upload request. This is the size of all the parts in the upload except the last part, which may be smaller than this size.

creationDate :: Maybe Text

The UTC time at which the multipart upload was initiated.

vaultARN :: Maybe Text

The Amazon Resource Name (ARN) of the vault that contains the archive.

archiveDescription :: Maybe Text

The description of the archive that was specified in the Initiate Multipart Upload request.

multipartUploadId :: Maybe Text

The ID of a multipart upload.

newUploadListElement :: UploadListElement

Create a value of UploadListElement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:partSizeInBytes:UploadListElement', uploadListElement_partSizeInBytes - The part size, in bytes, specified in the Initiate Multipart Upload request. This is the size of all the parts in the upload except the last part, which may be smaller than this size.

$sel:creationDate:UploadListElement', uploadListElement_creationDate - The UTC time at which the multipart upload was initiated.

$sel:vaultARN:UploadListElement', uploadListElement_vaultARN - The Amazon Resource Name (ARN) of the vault that contains the archive.

$sel:archiveDescription:UploadListElement', uploadListElement_archiveDescription - The description of the archive that was specified in the Initiate Multipart Upload request.

$sel:multipartUploadId:UploadListElement', uploadListElement_multipartUploadId - The ID of a multipart upload.

uploadListElement_partSizeInBytes :: Lens' UploadListElement (Maybe Integer)

The part size, in bytes, specified in the Initiate Multipart Upload request. This is the size of all the parts in the upload except the last part, which may be smaller than this size.

uploadListElement_creationDate :: Lens' UploadListElement (Maybe Text)

The UTC time at which the multipart upload was initiated.

uploadListElement_vaultARN :: Lens' UploadListElement (Maybe Text)

The Amazon Resource Name (ARN) of the vault that contains the archive.

uploadListElement_archiveDescription :: Lens' UploadListElement (Maybe Text)

The description of the archive that was specified in the Initiate Multipart Upload request.

uploadListElement_multipartUploadId :: Lens' UploadListElement (Maybe Text)

The ID of a multipart upload.