Network.AWS.Glacier.UploadMultipartPart

Description

This operation uploads a part of an archive. You can upload archive parts in any order. You can also upload them in parallel. You can upload up to 10,000 parts for a multipart upload.

Amazon Glacier rejects your upload part request if any of the following conditions is true:

SHA256 tree hash does not matchTo ensure that part data is not corrupted in transmission, you compute a SHA256 tree hash of the part and include it in your request. Upon receiving the part data, Amazon S3 Glacier also computes a SHA256 tree hash. If these hash values don't match, the operation fails. For information about computing a SHA256 tree hash, see Computing Checksums.

Part size does not matchThe size of each part except the last must match the size specified in the corresponding InitiateMultipartUpload request. The size of the last part must be the same size as, or smaller than, the specified size.

If you upload a part whose size is smaller than the part size you specified in your initiate multipart upload request and that part is not the last part, then the upload part request will succeed. However, the subsequent Complete Multipart Upload request will fail.

Range does not alignThe byte range value in the request does not align with the part size specified in the corresponding initiate request. For example, if you specify a part size of 4194304 bytes (4 MB), then 0 to 4194303 bytes (4 MB - 1) and 4194304 (4 MB) to 8388607 (8 MB - 1) are valid part ranges. However, if you set a range value of 2 MB to 6 MB, the range does not align with the part size and the upload will fail.

This operation is idempotent. If you upload the same part multiple times, the data included in the most recent request overwrites the previously uploaded data.

An AWS account has full permission to perform all operations (actions). However, AWS Identity and Access Management (IAM) users don't have any permissions by default. You must grant them explicit permission to perform specific actions. For more information, see Access Control Using AWS Identity and Access Management (IAM).

For conceptual information and underlying REST API, see Uploading Large Archives in Parts (Multipart Upload) and Upload Part in the Amazon Glacier Developer Guide.

Creating a Request

data UploadMultipartPart

Provides options to upload a part of an archive in a multipart upload operation.

See: newUploadMultipartPart smart constructor.

Constructors

UploadMultipartPart'

Fields

accountId :: Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

vaultName :: Text

The name of the vault.

uploadId :: Text

The upload ID of the multipart upload.

range :: Text

Identifies the range of bytes in the assembled archive that will be uploaded in this part. Amazon S3 Glacier uses this information to assemble the archive in the proper sequence. The format of this header follows RFC 2616. An example header is Content-Range:bytes 0-4194303/*.

checksum :: Text

The SHA256 tree hash of the data being uploaded.

body :: HashedBody

The data to upload.

newUploadMultipartPart

Arguments

:: Text
$sel:accountId:UploadMultipartPart'

-> Text
$sel:vaultName:UploadMultipartPart'

-> Text
$sel:uploadId:UploadMultipartPart'

-> Text
$sel:range:UploadMultipartPart'

-> Text
$sel:checksum:UploadMultipartPart'

-> HashedBody
$sel:body:UploadMultipartPart'

-> UploadMultipartPart

Create a value of UploadMultipartPart with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:UploadMultipartPart', uploadMultipartPart_accountId - The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

$sel:vaultName:UploadMultipartPart', uploadMultipartPart_vaultName - The name of the vault.

$sel:uploadId:UploadMultipartPart', uploadMultipartPart_uploadId - The upload ID of the multipart upload.

$sel:range:UploadMultipartPart', uploadMultipartPart_range - Identifies the range of bytes in the assembled archive that will be uploaded in this part. Amazon S3 Glacier uses this information to assemble the archive in the proper sequence. The format of this header follows RFC 2616. An example header is Content-Range:bytes 0-4194303/*.

$sel:checksum:UploadMultipartPart', uploadMultipartPart_checksum - The SHA256 tree hash of the data being uploaded.

$sel:body:UploadMultipartPart', uploadMultipartPart_body - The data to upload.

Request Lenses

uploadMultipartPart_accountId :: Lens' UploadMultipartPart Text

The AccountId value is the AWS account ID of the account that owns the vault. You can either specify an AWS account ID or optionally a single '-' (hyphen), in which case Amazon S3 Glacier uses the AWS account ID associated with the credentials used to sign the request. If you use an account ID, do not include any hyphens ('-') in the ID.

uploadMultipartPart_vaultName :: Lens' UploadMultipartPart Text

The name of the vault.

uploadMultipartPart_uploadId :: Lens' UploadMultipartPart Text

The upload ID of the multipart upload.

uploadMultipartPart_range :: Lens' UploadMultipartPart Text

Identifies the range of bytes in the assembled archive that will be uploaded in this part. Amazon S3 Glacier uses this information to assemble the archive in the proper sequence. The format of this header follows RFC 2616. An example header is Content-Range:bytes 0-4194303/*.

uploadMultipartPart_checksum :: Lens' UploadMultipartPart Text

The SHA256 tree hash of the data being uploaded.

uploadMultipartPart_body :: Lens' UploadMultipartPart HashedBody

The data to upload.

Destructuring the Response

data UploadMultipartPartResponse

Contains the Amazon S3 Glacier response to your request.

See: newUploadMultipartPartResponse smart constructor.

Constructors

UploadMultipartPartResponse'

Fields

checksum :: Maybe Text

The SHA256 tree hash that Amazon S3 Glacier computed for the uploaded part.

httpStatus :: Int

The response's http status code.

newUploadMultipartPartResponse

Arguments

:: Int
$sel:httpStatus:UploadMultipartPartResponse'

-> UploadMultipartPartResponse

Create a value of UploadMultipartPartResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checksum:UploadMultipartPart', uploadMultipartPartResponse_checksum - The SHA256 tree hash that Amazon S3 Glacier computed for the uploaded part.

$sel:httpStatus:UploadMultipartPartResponse', uploadMultipartPartResponse_httpStatus - The response's http status code.

Response Lenses

uploadMultipartPartResponse_checksum :: Lens' UploadMultipartPartResponse (Maybe Text)

The SHA256 tree hash that Amazon S3 Glacier computed for the uploaded part.

uploadMultipartPartResponse_httpStatus :: Lens' UploadMultipartPartResponse Int

The response's http status code.