Network.AWS.GlobalAccelerator.CreateCustomRoutingEndpointGroup

Description

Create an endpoint group for the specified listener for a custom routing accelerator. An endpoint group is a collection of endpoints in one AWS Region.

Creating a Request

data CreateCustomRoutingEndpointGroup

See: newCreateCustomRoutingEndpointGroup smart constructor.

Constructors

CreateCustomRoutingEndpointGroup'

Fields

listenerArn :: Text

The Amazon Resource Name (ARN) of the listener for a custom routing endpoint.

endpointGroupRegion :: Text

The AWS Region where the endpoint group is located. A listener can have only one endpoint group in a specific Region.

destinationConfigurations :: NonEmpty CustomRoutingDestinationConfiguration

Sets the port range and protocol for all endpoints (virtual private cloud subnets) in a custom routing endpoint group to accept client traffic on.

idempotencyToken :: Text

A unique, case-sensitive identifier that you provide to ensure the idempotency—that is, the uniqueness—of the request.

newCreateCustomRoutingEndpointGroup

Arguments

:: Text
$sel:listenerArn:CreateCustomRoutingEndpointGroup'

-> Text
$sel:endpointGroupRegion:CreateCustomRoutingEndpointGroup'

-> NonEmpty CustomRoutingDestinationConfiguration
$sel:destinationConfigurations:CreateCustomRoutingEndpointGroup'

-> Text
$sel:idempotencyToken:CreateCustomRoutingEndpointGroup'

-> CreateCustomRoutingEndpointGroup

Create a value of CreateCustomRoutingEndpointGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:listenerArn:CreateCustomRoutingEndpointGroup', createCustomRoutingEndpointGroup_listenerArn - The Amazon Resource Name (ARN) of the listener for a custom routing endpoint.

$sel:endpointGroupRegion:CreateCustomRoutingEndpointGroup', createCustomRoutingEndpointGroup_endpointGroupRegion - The AWS Region where the endpoint group is located. A listener can have only one endpoint group in a specific Region.

$sel:destinationConfigurations:CreateCustomRoutingEndpointGroup', createCustomRoutingEndpointGroup_destinationConfigurations - Sets the port range and protocol for all endpoints (virtual private cloud subnets) in a custom routing endpoint group to accept client traffic on.

$sel:idempotencyToken:CreateCustomRoutingEndpointGroup', createCustomRoutingEndpointGroup_idempotencyToken - A unique, case-sensitive identifier that you provide to ensure the idempotency—that is, the uniqueness—of the request.

Request Lenses

createCustomRoutingEndpointGroup_listenerArn :: Lens' CreateCustomRoutingEndpointGroup Text

The Amazon Resource Name (ARN) of the listener for a custom routing endpoint.

createCustomRoutingEndpointGroup_endpointGroupRegion :: Lens' CreateCustomRoutingEndpointGroup Text

The AWS Region where the endpoint group is located. A listener can have only one endpoint group in a specific Region.

createCustomRoutingEndpointGroup_destinationConfigurations :: Lens' CreateCustomRoutingEndpointGroup (NonEmpty CustomRoutingDestinationConfiguration)

Sets the port range and protocol for all endpoints (virtual private cloud subnets) in a custom routing endpoint group to accept client traffic on.

createCustomRoutingEndpointGroup_idempotencyToken :: Lens' CreateCustomRoutingEndpointGroup Text

A unique, case-sensitive identifier that you provide to ensure the idempotency—that is, the uniqueness—of the request.

Destructuring the Response

data CreateCustomRoutingEndpointGroupResponse

See: newCreateCustomRoutingEndpointGroupResponse smart constructor.

Constructors

CreateCustomRoutingEndpointGroupResponse'

Fields

endpointGroup :: Maybe CustomRoutingEndpointGroup

The information about the endpoint group created for a custom routing accelerator.

httpStatus :: Int

The response's http status code.

newCreateCustomRoutingEndpointGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateCustomRoutingEndpointGroupResponse'

-> CreateCustomRoutingEndpointGroupResponse

Create a value of CreateCustomRoutingEndpointGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpointGroup:CreateCustomRoutingEndpointGroupResponse', createCustomRoutingEndpointGroupResponse_endpointGroup - The information about the endpoint group created for a custom routing accelerator.

$sel:httpStatus:CreateCustomRoutingEndpointGroupResponse', createCustomRoutingEndpointGroupResponse_httpStatus - The response's http status code.

Response Lenses

createCustomRoutingEndpointGroupResponse_endpointGroup :: Lens' CreateCustomRoutingEndpointGroupResponse (Maybe CustomRoutingEndpointGroup)

The information about the endpoint group created for a custom routing accelerator.

createCustomRoutingEndpointGroupResponse_httpStatus :: Lens' CreateCustomRoutingEndpointGroupResponse Int

The response's http status code.