Network.AWS.GlobalAccelerator.DescribeCustomRoutingEndpointGroup

Description

Describe an endpoint group for a custom routing accelerator.

Creating a Request

data DescribeCustomRoutingEndpointGroup

See: newDescribeCustomRoutingEndpointGroup smart constructor.

Constructors

DescribeCustomRoutingEndpointGroup'

Fields

endpointGroupArn :: Text

The Amazon Resource Name (ARN) of the endpoint group to describe.

newDescribeCustomRoutingEndpointGroup

Arguments

:: Text
$sel:endpointGroupArn:DescribeCustomRoutingEndpointGroup'

-> DescribeCustomRoutingEndpointGroup

Create a value of DescribeCustomRoutingEndpointGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpointGroupArn:DescribeCustomRoutingEndpointGroup', describeCustomRoutingEndpointGroup_endpointGroupArn - The Amazon Resource Name (ARN) of the endpoint group to describe.

Request Lenses

describeCustomRoutingEndpointGroup_endpointGroupArn :: Lens' DescribeCustomRoutingEndpointGroup Text

The Amazon Resource Name (ARN) of the endpoint group to describe.

Destructuring the Response

data DescribeCustomRoutingEndpointGroupResponse

See: newDescribeCustomRoutingEndpointGroupResponse smart constructor.

Constructors

DescribeCustomRoutingEndpointGroupResponse'

Fields

endpointGroup :: Maybe CustomRoutingEndpointGroup

The description of an endpoint group for a custom routing accelerator.

httpStatus :: Int

The response's http status code.

newDescribeCustomRoutingEndpointGroupResponse

Arguments

:: Int
$sel:httpStatus:DescribeCustomRoutingEndpointGroupResponse'

-> DescribeCustomRoutingEndpointGroupResponse

Create a value of DescribeCustomRoutingEndpointGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpointGroup:DescribeCustomRoutingEndpointGroupResponse', describeCustomRoutingEndpointGroupResponse_endpointGroup - The description of an endpoint group for a custom routing accelerator.

$sel:httpStatus:DescribeCustomRoutingEndpointGroupResponse', describeCustomRoutingEndpointGroupResponse_httpStatus - The response's http status code.

Response Lenses

describeCustomRoutingEndpointGroupResponse_endpointGroup :: Lens' DescribeCustomRoutingEndpointGroupResponse (Maybe CustomRoutingEndpointGroup)

The description of an endpoint group for a custom routing accelerator.

describeCustomRoutingEndpointGroupResponse_httpStatus :: Lens' DescribeCustomRoutingEndpointGroupResponse Int

The response's http status code.