Network.AWS.GlobalAccelerator.ListAccelerators

Description

List the accelerators for an AWS account.

This operation returns paginated results.

Creating a Request

data ListAccelerators

See: newListAccelerators smart constructor.

Constructors

ListAccelerators'

Fields

nextToken :: Maybe Text

The token for the next set of results. You receive this token from a previous call.

maxResults :: Maybe Natural

The number of Global Accelerator objects that you want to return with this call. The default value is 10.

newListAccelerators :: ListAccelerators

Create a value of ListAccelerators with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAccelerators', listAccelerators_nextToken - The token for the next set of results. You receive this token from a previous call.

$sel:maxResults:ListAccelerators', listAccelerators_maxResults - The number of Global Accelerator objects that you want to return with this call. The default value is 10.

Request Lenses

listAccelerators_nextToken :: Lens' ListAccelerators (Maybe Text)

The token for the next set of results. You receive this token from a previous call.

listAccelerators_maxResults :: Lens' ListAccelerators (Maybe Natural)

The number of Global Accelerator objects that you want to return with this call. The default value is 10.

Destructuring the Response

data ListAcceleratorsResponse

See: newListAcceleratorsResponse smart constructor.

Constructors

ListAcceleratorsResponse'

Fields

nextToken :: Maybe Text

The token for the next set of results. You receive this token from a previous call.

accelerators :: Maybe [Accelerator]

The list of accelerators for a customer account.

httpStatus :: Int

The response's http status code.

newListAcceleratorsResponse

Arguments

:: Int
$sel:httpStatus:ListAcceleratorsResponse'

-> ListAcceleratorsResponse

Create a value of ListAcceleratorsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAccelerators', listAcceleratorsResponse_nextToken - The token for the next set of results. You receive this token from a previous call.

$sel:accelerators:ListAcceleratorsResponse', listAcceleratorsResponse_accelerators - The list of accelerators for a customer account.

$sel:httpStatus:ListAcceleratorsResponse', listAcceleratorsResponse_httpStatus - The response's http status code.

Response Lenses

listAcceleratorsResponse_nextToken :: Lens' ListAcceleratorsResponse (Maybe Text)

The token for the next set of results. You receive this token from a previous call.

listAcceleratorsResponse_accelerators :: Lens' ListAcceleratorsResponse (Maybe [Accelerator])

The list of accelerators for a customer account.

listAcceleratorsResponse_httpStatus :: Lens' ListAcceleratorsResponse Int

The response's http status code.