Network.AWS.GlobalAccelerator.ListListeners

Description

List the listeners for an accelerator.

This operation returns paginated results.

Creating a Request

data ListListeners

See: newListListeners smart constructor.

Constructors

ListListeners'

Fields

nextToken :: Maybe Text

The token for the next set of results. You receive this token from a previous call.

maxResults :: Maybe Natural

The number of listener objects that you want to return with this call. The default value is 10.

acceleratorArn :: Text

The Amazon Resource Name (ARN) of the accelerator for which you want to list listener objects.

newListListeners

Arguments

:: Text
$sel:acceleratorArn:ListListeners'

-> ListListeners

Create a value of ListListeners with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListListeners', listListeners_nextToken - The token for the next set of results. You receive this token from a previous call.

$sel:maxResults:ListListeners', listListeners_maxResults - The number of listener objects that you want to return with this call. The default value is 10.

$sel:acceleratorArn:ListListeners', listListeners_acceleratorArn - The Amazon Resource Name (ARN) of the accelerator for which you want to list listener objects.

Request Lenses

listListeners_nextToken :: Lens' ListListeners (Maybe Text)

The token for the next set of results. You receive this token from a previous call.

listListeners_maxResults :: Lens' ListListeners (Maybe Natural)

The number of listener objects that you want to return with this call. The default value is 10.

listListeners_acceleratorArn :: Lens' ListListeners Text

The Amazon Resource Name (ARN) of the accelerator for which you want to list listener objects.

Destructuring the Response

data ListListenersResponse

See: newListListenersResponse smart constructor.

Constructors

ListListenersResponse'

Fields

nextToken :: Maybe Text

The token for the next set of results. You receive this token from a previous call.

listeners :: Maybe [Listener]

The list of listeners for an accelerator.

httpStatus :: Int

The response's http status code.

newListListenersResponse

Arguments

:: Int
$sel:httpStatus:ListListenersResponse'

-> ListListenersResponse

Create a value of ListListenersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListListeners', listListenersResponse_nextToken - The token for the next set of results. You receive this token from a previous call.

$sel:listeners:ListListenersResponse', listListenersResponse_listeners - The list of listeners for an accelerator.

$sel:httpStatus:ListListenersResponse', listListenersResponse_httpStatus - The response's http status code.

Response Lenses

listListenersResponse_nextToken :: Lens' ListListenersResponse (Maybe Text)

The token for the next set of results. You receive this token from a previous call.

listListenersResponse_listeners :: Lens' ListListenersResponse (Maybe [Listener])

The list of listeners for an accelerator.

listListenersResponse_httpStatus :: Lens' ListListenersResponse Int

The response's http status code.