Network.AWS.GlobalAccelerator.Types.Accelerator

Description

Documentation

data Accelerator

An accelerator is a complex type that includes one or more listeners that process inbound connections and then direct traffic to one or more endpoint groups, each of which includes endpoints, such as load balancers.

See: newAccelerator smart constructor.

Constructors

Accelerator'

Fields

status :: Maybe AcceleratorStatus

Describes the deployment status of the accelerator.

acceleratorArn :: Maybe Text

The Amazon Resource Name (ARN) of the accelerator.

enabled :: Maybe Bool

Indicates whether the accelerator is enabled. The value is true or false. The default value is true.

If the value is set to true, the accelerator cannot be deleted. If set to false, accelerator can be deleted.

createdTime :: Maybe POSIX

The date and time that the accelerator was created.

lastModifiedTime :: Maybe POSIX

The date and time that the accelerator was last modified.

ipAddressType :: Maybe IpAddressType

The value for the address type must be IPv4.

name :: Maybe Text

The name of the accelerator. The name must contain only alphanumeric characters or hyphens (-), and must not begin or end with a hyphen.

ipSets :: Maybe [IpSet]

The static IP addresses that Global Accelerator associates with the accelerator.

dnsName :: Maybe Text

The Domain Name System (DNS) name that Global Accelerator creates that points to your accelerator's static IP addresses.

The naming convention for the DNS name is the following: A lowercase letter a, followed by a 16-bit random hex string, followed by .awsglobalaccelerator.com. For example: a1234567890abcdef.awsglobalaccelerator.com.

For more information about the default DNS name, see Support for DNS Addressing in Global Accelerator in the AWS Global Accelerator Developer Guide.

newAccelerator :: Accelerator

Create a value of Accelerator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Accelerator', accelerator_status - Describes the deployment status of the accelerator.

$sel:acceleratorArn:Accelerator', accelerator_acceleratorArn - The Amazon Resource Name (ARN) of the accelerator.

$sel:enabled:Accelerator', accelerator_enabled - Indicates whether the accelerator is enabled. The value is true or false. The default value is true.

If the value is set to true, the accelerator cannot be deleted. If set to false, accelerator can be deleted.

$sel:createdTime:Accelerator', accelerator_createdTime - The date and time that the accelerator was created.

$sel:lastModifiedTime:Accelerator', accelerator_lastModifiedTime - The date and time that the accelerator was last modified.

$sel:ipAddressType:Accelerator', accelerator_ipAddressType - The value for the address type must be IPv4.

$sel:name:Accelerator', accelerator_name - The name of the accelerator. The name must contain only alphanumeric characters or hyphens (-), and must not begin or end with a hyphen.

$sel:ipSets:Accelerator', accelerator_ipSets - The static IP addresses that Global Accelerator associates with the accelerator.

$sel:dnsName:Accelerator', accelerator_dnsName - The Domain Name System (DNS) name that Global Accelerator creates that points to your accelerator's static IP addresses.

The naming convention for the DNS name is the following: A lowercase letter a, followed by a 16-bit random hex string, followed by .awsglobalaccelerator.com. For example: a1234567890abcdef.awsglobalaccelerator.com.

For more information about the default DNS name, see Support for DNS Addressing in Global Accelerator in the AWS Global Accelerator Developer Guide.

accelerator_status :: Lens' Accelerator (Maybe AcceleratorStatus)

Describes the deployment status of the accelerator.

accelerator_acceleratorArn :: Lens' Accelerator (Maybe Text)

The Amazon Resource Name (ARN) of the accelerator.

accelerator_enabled :: Lens' Accelerator (Maybe Bool)

Indicates whether the accelerator is enabled. The value is true or false. The default value is true.

If the value is set to true, the accelerator cannot be deleted. If set to false, accelerator can be deleted.

accelerator_createdTime :: Lens' Accelerator (Maybe UTCTime)

The date and time that the accelerator was created.

accelerator_lastModifiedTime :: Lens' Accelerator (Maybe UTCTime)

The date and time that the accelerator was last modified.

accelerator_ipAddressType :: Lens' Accelerator (Maybe IpAddressType)

The value for the address type must be IPv4.

accelerator_name :: Lens' Accelerator (Maybe Text)

The name of the accelerator. The name must contain only alphanumeric characters or hyphens (-), and must not begin or end with a hyphen.

accelerator_ipSets :: Lens' Accelerator (Maybe [IpSet])

The static IP addresses that Global Accelerator associates with the accelerator.

accelerator_dnsName :: Lens' Accelerator (Maybe Text)

The Domain Name System (DNS) name that Global Accelerator creates that points to your accelerator's static IP addresses.

The naming convention for the DNS name is the following: A lowercase letter a, followed by a 16-bit random hex string, followed by .awsglobalaccelerator.com. For example: a1234567890abcdef.awsglobalaccelerator.com.

For more information about the default DNS name, see Support for DNS Addressing in Global Accelerator in the AWS Global Accelerator Developer Guide.