Network.AWS.GlobalAccelerator.Types.CustomRoutingAcceleratorAttributes

Description

Documentation

data CustomRoutingAcceleratorAttributes

Attributes of a custom routing accelerator.

See: newCustomRoutingAcceleratorAttributes smart constructor.

Constructors

CustomRoutingAcceleratorAttributes'

Fields

flowLogsS3Prefix :: Maybe Text

The prefix for the location in the Amazon S3 bucket for the flow logs. Attribute is required if FlowLogsEnabled is true.

If you don’t specify a prefix, the flow logs are stored in the root of the bucket. If you specify slash (/) for the S3 bucket prefix, the log file bucket folder structure will include a double slash (//), like the following:

DOC-EXAMPLE-BUCKET//AWSLogs/aws_account_id

flowLogsEnabled :: Maybe Bool

Indicates whether flow logs are enabled. The default value is false. If the value is true, FlowLogsS3Bucket and FlowLogsS3Prefix must be specified.

For more information, see Flow Logs in the AWS Global Accelerator Developer Guide.

flowLogsS3Bucket :: Maybe Text

The name of the Amazon S3 bucket for the flow logs. Attribute is required if FlowLogsEnabled is true. The bucket must exist and have a bucket policy that grants AWS Global Accelerator permission to write to the bucket.

newCustomRoutingAcceleratorAttributes :: CustomRoutingAcceleratorAttributes

Create a value of CustomRoutingAcceleratorAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flowLogsS3Prefix:CustomRoutingAcceleratorAttributes', customRoutingAcceleratorAttributes_flowLogsS3Prefix - The prefix for the location in the Amazon S3 bucket for the flow logs. Attribute is required if FlowLogsEnabled is true.

If you don’t specify a prefix, the flow logs are stored in the root of the bucket. If you specify slash (/) for the S3 bucket prefix, the log file bucket folder structure will include a double slash (//), like the following:

DOC-EXAMPLE-BUCKET//AWSLogs/aws_account_id

$sel:flowLogsEnabled:CustomRoutingAcceleratorAttributes', customRoutingAcceleratorAttributes_flowLogsEnabled - Indicates whether flow logs are enabled. The default value is false. If the value is true, FlowLogsS3Bucket and FlowLogsS3Prefix must be specified.

For more information, see Flow Logs in the AWS Global Accelerator Developer Guide.

$sel:flowLogsS3Bucket:CustomRoutingAcceleratorAttributes', customRoutingAcceleratorAttributes_flowLogsS3Bucket - The name of the Amazon S3 bucket for the flow logs. Attribute is required if FlowLogsEnabled is true. The bucket must exist and have a bucket policy that grants AWS Global Accelerator permission to write to the bucket.

customRoutingAcceleratorAttributes_flowLogsS3Prefix :: Lens' CustomRoutingAcceleratorAttributes (Maybe Text)

The prefix for the location in the Amazon S3 bucket for the flow logs. Attribute is required if FlowLogsEnabled is true.

If you don’t specify a prefix, the flow logs are stored in the root of the bucket. If you specify slash (/) for the S3 bucket prefix, the log file bucket folder structure will include a double slash (//), like the following:

DOC-EXAMPLE-BUCKET//AWSLogs/aws_account_id

customRoutingAcceleratorAttributes_flowLogsEnabled :: Lens' CustomRoutingAcceleratorAttributes (Maybe Bool)

Indicates whether flow logs are enabled. The default value is false. If the value is true, FlowLogsS3Bucket and FlowLogsS3Prefix must be specified.

For more information, see Flow Logs in the AWS Global Accelerator Developer Guide.

customRoutingAcceleratorAttributes_flowLogsS3Bucket :: Lens' CustomRoutingAcceleratorAttributes (Maybe Text)

The name of the Amazon S3 bucket for the flow logs. Attribute is required if FlowLogsEnabled is true. The bucket must exist and have a bucket policy that grants AWS Global Accelerator permission to write to the bucket.