Network.AWS.GlobalAccelerator.Types.CustomRoutingDestinationConfiguration

Description

Documentation

data CustomRoutingDestinationConfiguration

For a custom routing accelerator, sets the port range and protocol for all endpoints (virtual private cloud subnets) in an endpoint group to accept client traffic on.

See: newCustomRoutingDestinationConfiguration smart constructor.

Constructors

CustomRoutingDestinationConfiguration'

Fields

fromPort :: Natural

The first port, inclusive, in the range of ports for the endpoint group that is associated with a custom routing accelerator.

toPort :: Natural

The last port, inclusive, in the range of ports for the endpoint group that is associated with a custom routing accelerator.

protocols :: NonEmpty CustomRoutingProtocol

The protocol for the endpoint group that is associated with a custom routing accelerator. The protocol can be either TCP or UDP.

newCustomRoutingDestinationConfiguration

Arguments

:: Natural
$sel:fromPort:CustomRoutingDestinationConfiguration'

-> Natural
$sel:toPort:CustomRoutingDestinationConfiguration'

-> NonEmpty CustomRoutingProtocol
$sel:protocols:CustomRoutingDestinationConfiguration'

-> CustomRoutingDestinationConfiguration

Create a value of CustomRoutingDestinationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fromPort:CustomRoutingDestinationConfiguration', customRoutingDestinationConfiguration_fromPort - The first port, inclusive, in the range of ports for the endpoint group that is associated with a custom routing accelerator.

$sel:toPort:CustomRoutingDestinationConfiguration', customRoutingDestinationConfiguration_toPort - The last port, inclusive, in the range of ports for the endpoint group that is associated with a custom routing accelerator.

$sel:protocols:CustomRoutingDestinationConfiguration', customRoutingDestinationConfiguration_protocols - The protocol for the endpoint group that is associated with a custom routing accelerator. The protocol can be either TCP or UDP.

customRoutingDestinationConfiguration_fromPort :: Lens' CustomRoutingDestinationConfiguration Natural

The first port, inclusive, in the range of ports for the endpoint group that is associated with a custom routing accelerator.

customRoutingDestinationConfiguration_toPort :: Lens' CustomRoutingDestinationConfiguration Natural

The last port, inclusive, in the range of ports for the endpoint group that is associated with a custom routing accelerator.

customRoutingDestinationConfiguration_protocols :: Lens' CustomRoutingDestinationConfiguration (NonEmpty CustomRoutingProtocol)

The protocol for the endpoint group that is associated with a custom routing accelerator. The protocol can be either TCP or UDP.