Network.AWS.GlobalAccelerator.Types.PortMapping

Description

Documentation

data PortMapping

Returns the ports and associated IP addresses and ports of Amazon EC2 instances in your virtual private cloud (VPC) subnets. Custom routing is a port mapping protocol in AWS Global Accelerator that statically associates port ranges with VPC subnets, which allows Global Accelerator to route to specific instances and ports within one or more subnets.

See: newPortMapping smart constructor.

Constructors

PortMapping'

Fields

destinationSocketAddress :: Maybe SocketAddress

The EC2 instance IP address and port number in the virtual private cloud (VPC) subnet.

protocols :: Maybe (NonEmpty CustomRoutingProtocol)

The protocols supported by the endpoint group.

endpointGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of the endpoint group.

endpointId :: Maybe Text

The IP address of the VPC subnet (the subnet ID).

destinationTrafficState :: Maybe CustomRoutingDestinationTrafficState

Indicates whether or not a port mapping destination can receive traffic. The value is either ALLOW, if traffic is allowed to the destination, or DENY, if traffic is not allowed to the destination.

acceleratorPort :: Maybe Natural

The accelerator port.

newPortMapping :: PortMapping

Create a value of PortMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationSocketAddress:PortMapping', portMapping_destinationSocketAddress - The EC2 instance IP address and port number in the virtual private cloud (VPC) subnet.

$sel:protocols:PortMapping', portMapping_protocols - The protocols supported by the endpoint group.

$sel:endpointGroupArn:PortMapping', portMapping_endpointGroupArn - The Amazon Resource Name (ARN) of the endpoint group.

$sel:endpointId:PortMapping', portMapping_endpointId - The IP address of the VPC subnet (the subnet ID).

$sel:destinationTrafficState:PortMapping', portMapping_destinationTrafficState - Indicates whether or not a port mapping destination can receive traffic. The value is either ALLOW, if traffic is allowed to the destination, or DENY, if traffic is not allowed to the destination.

$sel:acceleratorPort:PortMapping', portMapping_acceleratorPort - The accelerator port.

portMapping_destinationSocketAddress :: Lens' PortMapping (Maybe SocketAddress)

The EC2 instance IP address and port number in the virtual private cloud (VPC) subnet.

portMapping_protocols :: Lens' PortMapping (Maybe (NonEmpty CustomRoutingProtocol))

The protocols supported by the endpoint group.

portMapping_endpointGroupArn :: Lens' PortMapping (Maybe Text)

The Amazon Resource Name (ARN) of the endpoint group.

portMapping_endpointId :: Lens' PortMapping (Maybe Text)

The IP address of the VPC subnet (the subnet ID).

portMapping_destinationTrafficState :: Lens' PortMapping (Maybe CustomRoutingDestinationTrafficState)

Indicates whether or not a port mapping destination can receive traffic. The value is either ALLOW, if traffic is allowed to the destination, or DENY, if traffic is not allowed to the destination.

portMapping_acceleratorPort :: Lens' PortMapping (Maybe Natural)

The accelerator port.