Network.AWS.GlobalAccelerator.UpdateCustomRoutingAccelerator

Description

Update a custom routing accelerator.

Creating a Request

data UpdateCustomRoutingAccelerator

See: newUpdateCustomRoutingAccelerator smart constructor.

Constructors

UpdateCustomRoutingAccelerator'

Fields

enabled :: Maybe Bool

Indicates whether an accelerator is enabled. The value is true or false. The default value is true.

If the value is set to true, the accelerator cannot be deleted. If set to false, the accelerator can be deleted.

ipAddressType :: Maybe IpAddressType

The value for the address type must be IPv4.

name :: Maybe Text

The name of the accelerator. The name can have a maximum of 32 characters, must contain only alphanumeric characters or hyphens (-), and must not begin or end with a hyphen.

acceleratorArn :: Text

The Amazon Resource Name (ARN) of the accelerator to update.

newUpdateCustomRoutingAccelerator

Arguments

:: Text
$sel:acceleratorArn:UpdateCustomRoutingAccelerator'

-> UpdateCustomRoutingAccelerator

Create a value of UpdateCustomRoutingAccelerator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:UpdateCustomRoutingAccelerator', updateCustomRoutingAccelerator_enabled - Indicates whether an accelerator is enabled. The value is true or false. The default value is true.

If the value is set to true, the accelerator cannot be deleted. If set to false, the accelerator can be deleted.

$sel:ipAddressType:UpdateCustomRoutingAccelerator', updateCustomRoutingAccelerator_ipAddressType - The value for the address type must be IPv4.

$sel:name:UpdateCustomRoutingAccelerator', updateCustomRoutingAccelerator_name - The name of the accelerator. The name can have a maximum of 32 characters, must contain only alphanumeric characters or hyphens (-), and must not begin or end with a hyphen.

$sel:acceleratorArn:UpdateCustomRoutingAccelerator', updateCustomRoutingAccelerator_acceleratorArn - The Amazon Resource Name (ARN) of the accelerator to update.

Request Lenses

updateCustomRoutingAccelerator_enabled :: Lens' UpdateCustomRoutingAccelerator (Maybe Bool)

Indicates whether an accelerator is enabled. The value is true or false. The default value is true.

If the value is set to true, the accelerator cannot be deleted. If set to false, the accelerator can be deleted.

updateCustomRoutingAccelerator_ipAddressType :: Lens' UpdateCustomRoutingAccelerator (Maybe IpAddressType)

The value for the address type must be IPv4.

updateCustomRoutingAccelerator_name :: Lens' UpdateCustomRoutingAccelerator (Maybe Text)

The name of the accelerator. The name can have a maximum of 32 characters, must contain only alphanumeric characters or hyphens (-), and must not begin or end with a hyphen.

updateCustomRoutingAccelerator_acceleratorArn :: Lens' UpdateCustomRoutingAccelerator Text

The Amazon Resource Name (ARN) of the accelerator to update.

Destructuring the Response

data UpdateCustomRoutingAcceleratorResponse

See: newUpdateCustomRoutingAcceleratorResponse smart constructor.

Constructors

UpdateCustomRoutingAcceleratorResponse'

Fields

accelerator :: Maybe CustomRoutingAccelerator

Information about the updated custom routing accelerator.

httpStatus :: Int

The response's http status code.

newUpdateCustomRoutingAcceleratorResponse

Arguments

:: Int
$sel:httpStatus:UpdateCustomRoutingAcceleratorResponse'

-> UpdateCustomRoutingAcceleratorResponse

Create a value of UpdateCustomRoutingAcceleratorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accelerator:UpdateCustomRoutingAcceleratorResponse', updateCustomRoutingAcceleratorResponse_accelerator - Information about the updated custom routing accelerator.

$sel:httpStatus:UpdateCustomRoutingAcceleratorResponse', updateCustomRoutingAcceleratorResponse_httpStatus - The response's http status code.

Response Lenses

updateCustomRoutingAcceleratorResponse_accelerator :: Lens' UpdateCustomRoutingAcceleratorResponse (Maybe CustomRoutingAccelerator)

Information about the updated custom routing accelerator.

updateCustomRoutingAcceleratorResponse_httpStatus :: Lens' UpdateCustomRoutingAcceleratorResponse Int

The response's http status code.