Network.AWS.GlobalAccelerator.UpdateCustomRoutingAcceleratorAttributes

Description

Update the attributes for a custom routing accelerator.

Creating a Request

data UpdateCustomRoutingAcceleratorAttributes

See: newUpdateCustomRoutingAcceleratorAttributes smart constructor.

Constructors

UpdateCustomRoutingAcceleratorAttributes'

Fields

flowLogsS3Prefix :: Maybe Text

Update the prefix for the location in the Amazon S3 bucket for the flow logs. Attribute is required if FlowLogsEnabled is true.

If you don’t specify a prefix, the flow logs are stored in the root of the bucket. If you specify slash (/) for the S3 bucket prefix, the log file bucket folder structure will include a double slash (//), like the following:

DOC-EXAMPLE-BUCKET//AWSLogs/aws_account_id

flowLogsEnabled :: Maybe Bool

Update whether flow logs are enabled. The default value is false. If the value is true, FlowLogsS3Bucket and FlowLogsS3Prefix must be specified.

For more information, see Flow Logs in the AWS Global Accelerator Developer Guide.

flowLogsS3Bucket :: Maybe Text

The name of the Amazon S3 bucket for the flow logs. Attribute is required if FlowLogsEnabled is true. The bucket must exist and have a bucket policy that grants AWS Global Accelerator permission to write to the bucket.

acceleratorArn :: Text

The Amazon Resource Name (ARN) of the custom routing accelerator to update attributes for.

newUpdateCustomRoutingAcceleratorAttributes

Arguments

:: Text
$sel:acceleratorArn:UpdateCustomRoutingAcceleratorAttributes'

-> UpdateCustomRoutingAcceleratorAttributes

Create a value of UpdateCustomRoutingAcceleratorAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flowLogsS3Prefix:UpdateCustomRoutingAcceleratorAttributes', updateCustomRoutingAcceleratorAttributes_flowLogsS3Prefix - Update the prefix for the location in the Amazon S3 bucket for the flow logs. Attribute is required if FlowLogsEnabled is true.

If you don’t specify a prefix, the flow logs are stored in the root of the bucket. If you specify slash (/) for the S3 bucket prefix, the log file bucket folder structure will include a double slash (//), like the following:

DOC-EXAMPLE-BUCKET//AWSLogs/aws_account_id

$sel:flowLogsEnabled:UpdateCustomRoutingAcceleratorAttributes', updateCustomRoutingAcceleratorAttributes_flowLogsEnabled - Update whether flow logs are enabled. The default value is false. If the value is true, FlowLogsS3Bucket and FlowLogsS3Prefix must be specified.

For more information, see Flow Logs in the AWS Global Accelerator Developer Guide.

$sel:flowLogsS3Bucket:UpdateCustomRoutingAcceleratorAttributes', updateCustomRoutingAcceleratorAttributes_flowLogsS3Bucket - The name of the Amazon S3 bucket for the flow logs. Attribute is required if FlowLogsEnabled is true. The bucket must exist and have a bucket policy that grants AWS Global Accelerator permission to write to the bucket.

$sel:acceleratorArn:UpdateCustomRoutingAcceleratorAttributes', updateCustomRoutingAcceleratorAttributes_acceleratorArn - The Amazon Resource Name (ARN) of the custom routing accelerator to update attributes for.

Request Lenses

updateCustomRoutingAcceleratorAttributes_flowLogsS3Prefix :: Lens' UpdateCustomRoutingAcceleratorAttributes (Maybe Text)

Update the prefix for the location in the Amazon S3 bucket for the flow logs. Attribute is required if FlowLogsEnabled is true.

If you don’t specify a prefix, the flow logs are stored in the root of the bucket. If you specify slash (/) for the S3 bucket prefix, the log file bucket folder structure will include a double slash (//), like the following:

DOC-EXAMPLE-BUCKET//AWSLogs/aws_account_id

updateCustomRoutingAcceleratorAttributes_flowLogsEnabled :: Lens' UpdateCustomRoutingAcceleratorAttributes (Maybe Bool)

Update whether flow logs are enabled. The default value is false. If the value is true, FlowLogsS3Bucket and FlowLogsS3Prefix must be specified.

For more information, see Flow Logs in the AWS Global Accelerator Developer Guide.

updateCustomRoutingAcceleratorAttributes_flowLogsS3Bucket :: Lens' UpdateCustomRoutingAcceleratorAttributes (Maybe Text)

The name of the Amazon S3 bucket for the flow logs. Attribute is required if FlowLogsEnabled is true. The bucket must exist and have a bucket policy that grants AWS Global Accelerator permission to write to the bucket.

updateCustomRoutingAcceleratorAttributes_acceleratorArn :: Lens' UpdateCustomRoutingAcceleratorAttributes Text

The Amazon Resource Name (ARN) of the custom routing accelerator to update attributes for.

Destructuring the Response

data UpdateCustomRoutingAcceleratorAttributesResponse

See: newUpdateCustomRoutingAcceleratorAttributesResponse smart constructor.

Constructors

UpdateCustomRoutingAcceleratorAttributesResponse'

Fields

acceleratorAttributes :: Maybe CustomRoutingAcceleratorAttributes

Updated custom routing accelerator.

httpStatus :: Int

The response's http status code.

newUpdateCustomRoutingAcceleratorAttributesResponse

Arguments

:: Int
$sel:httpStatus:UpdateCustomRoutingAcceleratorAttributesResponse'

-> UpdateCustomRoutingAcceleratorAttributesResponse

Create a value of UpdateCustomRoutingAcceleratorAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acceleratorAttributes:UpdateCustomRoutingAcceleratorAttributesResponse', updateCustomRoutingAcceleratorAttributesResponse_acceleratorAttributes - Updated custom routing accelerator.

$sel:httpStatus:UpdateCustomRoutingAcceleratorAttributesResponse', updateCustomRoutingAcceleratorAttributesResponse_httpStatus - The response's http status code.

Response Lenses

updateCustomRoutingAcceleratorAttributesResponse_acceleratorAttributes :: Lens' UpdateCustomRoutingAcceleratorAttributesResponse (Maybe CustomRoutingAcceleratorAttributes)

Updated custom routing accelerator.

updateCustomRoutingAcceleratorAttributesResponse_httpStatus :: Lens' UpdateCustomRoutingAcceleratorAttributesResponse Int

The response's http status code.