Network.AWS.GlobalAccelerator.UpdateEndpointGroup

Description

Update an endpoint group. A resource must be valid and active when you add it as an endpoint.

Creating a Request

data UpdateEndpointGroup

See: newUpdateEndpointGroup smart constructor.

Constructors

UpdateEndpointGroup'

Fields

thresholdCount :: Maybe Natural

The number of consecutive health checks required to set the state of a healthy endpoint to unhealthy, or to set an unhealthy endpoint to healthy. The default value is 3.

healthCheckPath :: Maybe Text

If the protocol is HTTP/S, then this specifies the path that is the destination for health check targets. The default value is slash (/).

healthCheckIntervalSeconds :: Maybe Natural

The time—10 seconds or 30 seconds—between each health check for an endpoint. The default value is 30.

healthCheckProtocol :: Maybe HealthCheckProtocol

The protocol that AWS Global Accelerator uses to check the health of endpoints that are part of this endpoint group. The default value is TCP.

trafficDialPercentage :: Maybe Double

The percentage of traffic to send to an AWS Region. Additional traffic is distributed to other endpoint groups for this listener.

Use this action to increase (dial up) or decrease (dial down) traffic to a specific Region. The percentage is applied to the traffic that would otherwise have been routed to the Region based on optimal routing.

The default value is 100.

endpointConfigurations :: Maybe [EndpointConfiguration]

The list of endpoint objects. A resource must be valid and active when you add it as an endpoint.

healthCheckPort :: Maybe Natural

The port that AWS Global Accelerator uses to check the health of endpoints that are part of this endpoint group. The default port is the listener port that this endpoint group is associated with. If the listener port is a list of ports, Global Accelerator uses the first port in the list.

portOverrides :: Maybe [PortOverride]

Override specific listener ports used to route traffic to endpoints that are part of this endpoint group. For example, you can create a port override in which the listener receives user traffic on ports 80 and 443, but your accelerator routes that traffic to ports 1080 and 1443, respectively, on the endpoints.

For more information, see Port overrides in the AWS Global Accelerator Developer Guide.

endpointGroupArn :: Text

The Amazon Resource Name (ARN) of the endpoint group.

newUpdateEndpointGroup

Arguments

:: Text
$sel:endpointGroupArn:UpdateEndpointGroup'

-> UpdateEndpointGroup

Create a value of UpdateEndpointGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thresholdCount:UpdateEndpointGroup', updateEndpointGroup_thresholdCount - The number of consecutive health checks required to set the state of a healthy endpoint to unhealthy, or to set an unhealthy endpoint to healthy. The default value is 3.

$sel:healthCheckPath:UpdateEndpointGroup', updateEndpointGroup_healthCheckPath - If the protocol is HTTP/S, then this specifies the path that is the destination for health check targets. The default value is slash (/).

$sel:healthCheckIntervalSeconds:UpdateEndpointGroup', updateEndpointGroup_healthCheckIntervalSeconds - The time—10 seconds or 30 seconds—between each health check for an endpoint. The default value is 30.

$sel:healthCheckProtocol:UpdateEndpointGroup', updateEndpointGroup_healthCheckProtocol - The protocol that AWS Global Accelerator uses to check the health of endpoints that are part of this endpoint group. The default value is TCP.

$sel:trafficDialPercentage:UpdateEndpointGroup', updateEndpointGroup_trafficDialPercentage - The percentage of traffic to send to an AWS Region. Additional traffic is distributed to other endpoint groups for this listener.

Use this action to increase (dial up) or decrease (dial down) traffic to a specific Region. The percentage is applied to the traffic that would otherwise have been routed to the Region based on optimal routing.

The default value is 100.

$sel:endpointConfigurations:UpdateEndpointGroup', updateEndpointGroup_endpointConfigurations - The list of endpoint objects. A resource must be valid and active when you add it as an endpoint.

$sel:healthCheckPort:UpdateEndpointGroup', updateEndpointGroup_healthCheckPort - The port that AWS Global Accelerator uses to check the health of endpoints that are part of this endpoint group. The default port is the listener port that this endpoint group is associated with. If the listener port is a list of ports, Global Accelerator uses the first port in the list.

$sel:portOverrides:UpdateEndpointGroup', updateEndpointGroup_portOverrides - Override specific listener ports used to route traffic to endpoints that are part of this endpoint group. For example, you can create a port override in which the listener receives user traffic on ports 80 and 443, but your accelerator routes that traffic to ports 1080 and 1443, respectively, on the endpoints.

For more information, see Port overrides in the AWS Global Accelerator Developer Guide.

$sel:endpointGroupArn:UpdateEndpointGroup', updateEndpointGroup_endpointGroupArn - The Amazon Resource Name (ARN) of the endpoint group.

Request Lenses

updateEndpointGroup_thresholdCount :: Lens' UpdateEndpointGroup (Maybe Natural)

The number of consecutive health checks required to set the state of a healthy endpoint to unhealthy, or to set an unhealthy endpoint to healthy. The default value is 3.

updateEndpointGroup_healthCheckPath :: Lens' UpdateEndpointGroup (Maybe Text)

If the protocol is HTTP/S, then this specifies the path that is the destination for health check targets. The default value is slash (/).

updateEndpointGroup_healthCheckIntervalSeconds :: Lens' UpdateEndpointGroup (Maybe Natural)

The time—10 seconds or 30 seconds—between each health check for an endpoint. The default value is 30.

updateEndpointGroup_healthCheckProtocol :: Lens' UpdateEndpointGroup (Maybe HealthCheckProtocol)

The protocol that AWS Global Accelerator uses to check the health of endpoints that are part of this endpoint group. The default value is TCP.

updateEndpointGroup_trafficDialPercentage :: Lens' UpdateEndpointGroup (Maybe Double)

The percentage of traffic to send to an AWS Region. Additional traffic is distributed to other endpoint groups for this listener.

Use this action to increase (dial up) or decrease (dial down) traffic to a specific Region. The percentage is applied to the traffic that would otherwise have been routed to the Region based on optimal routing.

The default value is 100.

updateEndpointGroup_endpointConfigurations :: Lens' UpdateEndpointGroup (Maybe [EndpointConfiguration])

The list of endpoint objects. A resource must be valid and active when you add it as an endpoint.

updateEndpointGroup_healthCheckPort :: Lens' UpdateEndpointGroup (Maybe Natural)

The port that AWS Global Accelerator uses to check the health of endpoints that are part of this endpoint group. The default port is the listener port that this endpoint group is associated with. If the listener port is a list of ports, Global Accelerator uses the first port in the list.

updateEndpointGroup_portOverrides :: Lens' UpdateEndpointGroup (Maybe [PortOverride])

Override specific listener ports used to route traffic to endpoints that are part of this endpoint group. For example, you can create a port override in which the listener receives user traffic on ports 80 and 443, but your accelerator routes that traffic to ports 1080 and 1443, respectively, on the endpoints.

For more information, see Port overrides in the AWS Global Accelerator Developer Guide.

updateEndpointGroup_endpointGroupArn :: Lens' UpdateEndpointGroup Text

The Amazon Resource Name (ARN) of the endpoint group.

Destructuring the Response

data UpdateEndpointGroupResponse

See: newUpdateEndpointGroupResponse smart constructor.

Constructors

UpdateEndpointGroupResponse'

Fields

endpointGroup :: Maybe EndpointGroup

The information about the endpoint group that was updated.

httpStatus :: Int

The response's http status code.

newUpdateEndpointGroupResponse

Arguments

:: Int
$sel:httpStatus:UpdateEndpointGroupResponse'

-> UpdateEndpointGroupResponse

Create a value of UpdateEndpointGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpointGroup:UpdateEndpointGroupResponse', updateEndpointGroupResponse_endpointGroup - The information about the endpoint group that was updated.

$sel:httpStatus:UpdateEndpointGroupResponse', updateEndpointGroupResponse_httpStatus - The response's http status code.

Response Lenses

updateEndpointGroupResponse_endpointGroup :: Lens' UpdateEndpointGroupResponse (Maybe EndpointGroup)

The information about the endpoint group that was updated.

updateEndpointGroupResponse_httpStatus :: Lens' UpdateEndpointGroupResponse Int

The response's http status code.