Network.AWS.GlobalAccelerator.UpdateListener

Description

Update a listener.

Creating a Request

data UpdateListener

See: newUpdateListener smart constructor.

Constructors

UpdateListener'

Fields

portRanges :: Maybe (NonEmpty PortRange)

The updated list of port ranges for the connections from clients to the accelerator.

protocol :: Maybe Protocol

The updated protocol for the connections from clients to the accelerator.

clientAffinity :: Maybe ClientAffinity

Client affinity lets you direct all requests from a user to the same endpoint, if you have stateful applications, regardless of the port and protocol of the client request. Client affinity gives you control over whether to always route each client to the same specific endpoint.

AWS Global Accelerator uses a consistent-flow hashing algorithm to choose the optimal endpoint for a connection. If client affinity is NONE, Global Accelerator uses the "five-tuple" (5-tuple) properties—source IP address, source port, destination IP address, destination port, and protocol—to select the hash value, and then chooses the best endpoint. However, with this setting, if someone uses different ports to connect to Global Accelerator, their connections might not be always routed to the same endpoint because the hash value changes.

If you want a given client to always be routed to the same endpoint, set client affinity to SOURCE_IP instead. When you use the SOURCE_IP setting, Global Accelerator uses the "two-tuple" (2-tuple) properties— source (client) IP address and destination IP address—to select the hash value.

The default value is NONE.

listenerArn :: Text

The Amazon Resource Name (ARN) of the listener to update.

newUpdateListener

Arguments

:: Text
$sel:listenerArn:UpdateListener'

-> UpdateListener

Create a value of UpdateListener with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:portRanges:UpdateListener', updateListener_portRanges - The updated list of port ranges for the connections from clients to the accelerator.

$sel:protocol:UpdateListener', updateListener_protocol - The updated protocol for the connections from clients to the accelerator.

$sel:clientAffinity:UpdateListener', updateListener_clientAffinity - Client affinity lets you direct all requests from a user to the same endpoint, if you have stateful applications, regardless of the port and protocol of the client request. Client affinity gives you control over whether to always route each client to the same specific endpoint.

AWS Global Accelerator uses a consistent-flow hashing algorithm to choose the optimal endpoint for a connection. If client affinity is NONE, Global Accelerator uses the "five-tuple" (5-tuple) properties—source IP address, source port, destination IP address, destination port, and protocol—to select the hash value, and then chooses the best endpoint. However, with this setting, if someone uses different ports to connect to Global Accelerator, their connections might not be always routed to the same endpoint because the hash value changes.

If you want a given client to always be routed to the same endpoint, set client affinity to SOURCE_IP instead. When you use the SOURCE_IP setting, Global Accelerator uses the "two-tuple" (2-tuple) properties— source (client) IP address and destination IP address—to select the hash value.

The default value is NONE.

$sel:listenerArn:UpdateListener', updateListener_listenerArn - The Amazon Resource Name (ARN) of the listener to update.

Request Lenses

updateListener_portRanges :: Lens' UpdateListener (Maybe (NonEmpty PortRange))

The updated list of port ranges for the connections from clients to the accelerator.

updateListener_protocol :: Lens' UpdateListener (Maybe Protocol)

The updated protocol for the connections from clients to the accelerator.

updateListener_clientAffinity :: Lens' UpdateListener (Maybe ClientAffinity)

Client affinity lets you direct all requests from a user to the same endpoint, if you have stateful applications, regardless of the port and protocol of the client request. Client affinity gives you control over whether to always route each client to the same specific endpoint.

AWS Global Accelerator uses a consistent-flow hashing algorithm to choose the optimal endpoint for a connection. If client affinity is NONE, Global Accelerator uses the "five-tuple" (5-tuple) properties—source IP address, source port, destination IP address, destination port, and protocol—to select the hash value, and then chooses the best endpoint. However, with this setting, if someone uses different ports to connect to Global Accelerator, their connections might not be always routed to the same endpoint because the hash value changes.

If you want a given client to always be routed to the same endpoint, set client affinity to SOURCE_IP instead. When you use the SOURCE_IP setting, Global Accelerator uses the "two-tuple" (2-tuple) properties— source (client) IP address and destination IP address—to select the hash value.

The default value is NONE.

updateListener_listenerArn :: Lens' UpdateListener Text

The Amazon Resource Name (ARN) of the listener to update.

Destructuring the Response

data UpdateListenerResponse

See: newUpdateListenerResponse smart constructor.

Constructors

UpdateListenerResponse'

Fields

listener :: Maybe Listener

Information for the updated listener.

httpStatus :: Int

The response's http status code.

newUpdateListenerResponse

Arguments

:: Int
$sel:httpStatus:UpdateListenerResponse'

-> UpdateListenerResponse

Create a value of UpdateListenerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:listener:UpdateListenerResponse', updateListenerResponse_listener - Information for the updated listener.

$sel:httpStatus:UpdateListenerResponse', updateListenerResponse_httpStatus - The response's http status code.

Response Lenses

updateListenerResponse_listener :: Lens' UpdateListenerResponse (Maybe Listener)

Information for the updated listener.

updateListenerResponse_httpStatus :: Lens' UpdateListenerResponse Int

The response's http status code.