Network.AWS.Glue.BatchDeleteConnection

Description

Deletes a list of connection definitions from the Data Catalog.

Creating a Request

data BatchDeleteConnection

See: newBatchDeleteConnection smart constructor.

Constructors

BatchDeleteConnection'

Fields

catalogId :: Maybe Text

The ID of the Data Catalog in which the connections reside. If none is provided, the Amazon Web Services account ID is used by default.

connectionNameList :: [Text]

A list of names of the connections to delete.

newBatchDeleteConnection :: BatchDeleteConnection

Create a value of BatchDeleteConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:BatchDeleteConnection', batchDeleteConnection_catalogId - The ID of the Data Catalog in which the connections reside. If none is provided, the Amazon Web Services account ID is used by default.

$sel:connectionNameList:BatchDeleteConnection', batchDeleteConnection_connectionNameList - A list of names of the connections to delete.

Request Lenses

batchDeleteConnection_catalogId :: Lens' BatchDeleteConnection (Maybe Text)

The ID of the Data Catalog in which the connections reside. If none is provided, the Amazon Web Services account ID is used by default.

batchDeleteConnection_connectionNameList :: Lens' BatchDeleteConnection [Text]

A list of names of the connections to delete.

Destructuring the Response

data BatchDeleteConnectionResponse

See: newBatchDeleteConnectionResponse smart constructor.

Constructors

BatchDeleteConnectionResponse'

Fields

succeeded :: Maybe [Text]

A list of names of the connection definitions that were successfully deleted.

errors :: Maybe (HashMap Text ErrorDetail)

A map of the names of connections that were not successfully deleted to error details.

httpStatus :: Int

The response's http status code.

newBatchDeleteConnectionResponse

Arguments

:: Int
$sel:httpStatus:BatchDeleteConnectionResponse'

-> BatchDeleteConnectionResponse

Create a value of BatchDeleteConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:succeeded:BatchDeleteConnectionResponse', batchDeleteConnectionResponse_succeeded - A list of names of the connection definitions that were successfully deleted.

$sel:errors:BatchDeleteConnectionResponse', batchDeleteConnectionResponse_errors - A map of the names of connections that were not successfully deleted to error details.

$sel:httpStatus:BatchDeleteConnectionResponse', batchDeleteConnectionResponse_httpStatus - The response's http status code.

Response Lenses

batchDeleteConnectionResponse_succeeded :: Lens' BatchDeleteConnectionResponse (Maybe [Text])

A list of names of the connection definitions that were successfully deleted.

batchDeleteConnectionResponse_errors :: Lens' BatchDeleteConnectionResponse (Maybe (HashMap Text ErrorDetail))

A map of the names of connections that were not successfully deleted to error details.

batchDeleteConnectionResponse_httpStatus :: Lens' BatchDeleteConnectionResponse Int

The response's http status code.