Network.AWS.Glue.BatchGetBlueprints

Description

Retrieves information about a list of blueprints.

Creating a Request

data BatchGetBlueprints

See: newBatchGetBlueprints smart constructor.

Constructors

BatchGetBlueprints'

Fields

includeParameterSpec :: Maybe Bool

Specifies whether or not to include the parameters, as a JSON string, for the blueprint in the response.

includeBlueprint :: Maybe Bool

Specifies whether or not to include the blueprint in the response.

names :: NonEmpty Text

A list of blueprint names.

newBatchGetBlueprints

Arguments

:: NonEmpty Text
$sel:names:BatchGetBlueprints'

-> BatchGetBlueprints

Create a value of BatchGetBlueprints with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeParameterSpec:BatchGetBlueprints', batchGetBlueprints_includeParameterSpec - Specifies whether or not to include the parameters, as a JSON string, for the blueprint in the response.

$sel:includeBlueprint:BatchGetBlueprints', batchGetBlueprints_includeBlueprint - Specifies whether or not to include the blueprint in the response.

$sel:names:BatchGetBlueprints', batchGetBlueprints_names - A list of blueprint names.

Request Lenses

batchGetBlueprints_includeParameterSpec :: Lens' BatchGetBlueprints (Maybe Bool)

Specifies whether or not to include the parameters, as a JSON string, for the blueprint in the response.

batchGetBlueprints_includeBlueprint :: Lens' BatchGetBlueprints (Maybe Bool)

Specifies whether or not to include the blueprint in the response.

batchGetBlueprints_names :: Lens' BatchGetBlueprints (NonEmpty Text)

A list of blueprint names.

Destructuring the Response

data BatchGetBlueprintsResponse

See: newBatchGetBlueprintsResponse smart constructor.

Constructors

BatchGetBlueprintsResponse'

Fields

blueprints :: Maybe [Blueprint]

Returns a list of blueprint as a Blueprints object.

missingBlueprints :: Maybe [Text]

Returns a list of BlueprintNames that were not found.

httpStatus :: Int

The response's http status code.

newBatchGetBlueprintsResponse

Arguments

:: Int
$sel:httpStatus:BatchGetBlueprintsResponse'

-> BatchGetBlueprintsResponse

Create a value of BatchGetBlueprintsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blueprints:BatchGetBlueprintsResponse', batchGetBlueprintsResponse_blueprints - Returns a list of blueprint as a Blueprints object.

$sel:missingBlueprints:BatchGetBlueprintsResponse', batchGetBlueprintsResponse_missingBlueprints - Returns a list of BlueprintNames that were not found.

$sel:httpStatus:BatchGetBlueprintsResponse', batchGetBlueprintsResponse_httpStatus - The response's http status code.

Response Lenses

batchGetBlueprintsResponse_blueprints :: Lens' BatchGetBlueprintsResponse (Maybe [Blueprint])

Returns a list of blueprint as a Blueprints object.

batchGetBlueprintsResponse_missingBlueprints :: Lens' BatchGetBlueprintsResponse (Maybe [Text])

Returns a list of BlueprintNames that were not found.

batchGetBlueprintsResponse_httpStatus :: Lens' BatchGetBlueprintsResponse Int

The response's http status code.