Network.AWS.Glue.CheckSchemaVersionValidity

Description

Validates the supplied schema. This call has no side effects, it simply validates using the supplied schema using DataFormat as the format. Since it does not take a schema set name, no compatibility checks are performed.

Creating a Request

data CheckSchemaVersionValidity

See: newCheckSchemaVersionValidity smart constructor.

Constructors

CheckSchemaVersionValidity'

Fields

dataFormat :: DataFormat

The data format of the schema definition. Currently AVRO and JSON are supported.

schemaDefinition :: Text

The definition of the schema that has to be validated.

newCheckSchemaVersionValidity

Arguments

:: DataFormat
$sel:dataFormat:CheckSchemaVersionValidity'

-> Text
$sel:schemaDefinition:CheckSchemaVersionValidity'

-> CheckSchemaVersionValidity

Create a value of CheckSchemaVersionValidity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataFormat:CheckSchemaVersionValidity', checkSchemaVersionValidity_dataFormat - The data format of the schema definition. Currently AVRO and JSON are supported.

$sel:schemaDefinition:CheckSchemaVersionValidity', checkSchemaVersionValidity_schemaDefinition - The definition of the schema that has to be validated.

Request Lenses

checkSchemaVersionValidity_dataFormat :: Lens' CheckSchemaVersionValidity DataFormat

The data format of the schema definition. Currently AVRO and JSON are supported.

checkSchemaVersionValidity_schemaDefinition :: Lens' CheckSchemaVersionValidity Text

The definition of the schema that has to be validated.

Destructuring the Response

data CheckSchemaVersionValidityResponse

See: newCheckSchemaVersionValidityResponse smart constructor.

Constructors

CheckSchemaVersionValidityResponse'

Fields

error :: Maybe Text

A validation failure error message.

valid :: Maybe Bool

Return true, if the schema is valid and false otherwise.

httpStatus :: Int

The response's http status code.

newCheckSchemaVersionValidityResponse

Arguments

:: Int
$sel:httpStatus:CheckSchemaVersionValidityResponse'

-> CheckSchemaVersionValidityResponse

Create a value of CheckSchemaVersionValidityResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:error:CheckSchemaVersionValidityResponse', checkSchemaVersionValidityResponse_error - A validation failure error message.

$sel:valid:CheckSchemaVersionValidityResponse', checkSchemaVersionValidityResponse_valid - Return true, if the schema is valid and false otherwise.

$sel:httpStatus:CheckSchemaVersionValidityResponse', checkSchemaVersionValidityResponse_httpStatus - The response's http status code.

Response Lenses

checkSchemaVersionValidityResponse_error :: Lens' CheckSchemaVersionValidityResponse (Maybe Text)

A validation failure error message.

checkSchemaVersionValidityResponse_valid :: Lens' CheckSchemaVersionValidityResponse (Maybe Bool)

Return true, if the schema is valid and false otherwise.

checkSchemaVersionValidityResponse_httpStatus :: Lens' CheckSchemaVersionValidityResponse Int

The response's http status code.