Network.AWS.Glue.CreateCrawler

Description

Creates a new crawler with specified targets, role, configuration, and optional schedule. At least one crawl target must be specified, in the s3Targets field, the jdbcTargets field, or the DynamoDBTargets field.

Creating a Request

data CreateCrawler

See: newCreateCrawler smart constructor.

Constructors

CreateCrawler'

Fields

schemaChangePolicy :: Maybe SchemaChangePolicy

The policy for the crawler's update and deletion behavior.

recrawlPolicy :: Maybe RecrawlPolicy

A policy that specifies whether to crawl the entire dataset again, or to crawl only folders that were added since the last crawler run.

classifiers :: Maybe [Text]

A list of custom classifiers that the user has registered. By default, all built-in classifiers are included in a crawl, but these custom classifiers always override the default classifiers for a given classification.

configuration :: Maybe Text

Crawler configuration information. This versioned JSON string allows users to specify aspects of a crawler's behavior. For more information, see Configuring a Crawler.

lineageConfiguration :: Maybe LineageConfiguration

Specifies data lineage configuration settings for the crawler.

tags :: Maybe (HashMap Text Text)

The tags to use with this crawler request. You may use tags to limit access to the crawler. For more information about tags in Glue, see Amazon Web Services Tags in Glue in the developer guide.

description :: Maybe Text

A description of the new crawler.

tablePrefix :: Maybe Text

The table prefix used for catalog tables that are created.

schedule :: Maybe Text

A cron expression used to specify the schedule (see Time-Based Schedules for Jobs and Crawlers. For example, to run something every day at 12:15 UTC, you would specify: cron(15 12 * * ? *).

crawlerSecurityConfiguration :: Maybe Text

The name of the SecurityConfiguration structure to be used by this crawler.

databaseName :: Maybe Text

The Glue database where results are written, such as: arn:aws:daylight:us-east-1::database/sometable/*.

name :: Text

Name of the new crawler.

role' :: Text

The IAM role or Amazon Resource Name (ARN) of an IAM role used by the new crawler to access customer resources.

targets :: CrawlerTargets

A list of collection of targets to crawl.

newCreateCrawler

Arguments

:: Text
$sel:name:CreateCrawler'

-> Text
$sel:role':CreateCrawler'

-> CrawlerTargets
$sel:targets:CreateCrawler'

-> CreateCrawler

Create a value of CreateCrawler with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaChangePolicy:CreateCrawler', createCrawler_schemaChangePolicy - The policy for the crawler's update and deletion behavior.

$sel:recrawlPolicy:CreateCrawler', createCrawler_recrawlPolicy - A policy that specifies whether to crawl the entire dataset again, or to crawl only folders that were added since the last crawler run.

$sel:classifiers:CreateCrawler', createCrawler_classifiers - A list of custom classifiers that the user has registered. By default, all built-in classifiers are included in a crawl, but these custom classifiers always override the default classifiers for a given classification.

$sel:configuration:CreateCrawler', createCrawler_configuration - Crawler configuration information. This versioned JSON string allows users to specify aspects of a crawler's behavior. For more information, see Configuring a Crawler.

$sel:lineageConfiguration:CreateCrawler', createCrawler_lineageConfiguration - Specifies data lineage configuration settings for the crawler.

$sel:tags:CreateCrawler', createCrawler_tags - The tags to use with this crawler request. You may use tags to limit access to the crawler. For more information about tags in Glue, see Amazon Web Services Tags in Glue in the developer guide.

$sel:description:CreateCrawler', createCrawler_description - A description of the new crawler.

$sel:tablePrefix:CreateCrawler', createCrawler_tablePrefix - The table prefix used for catalog tables that are created.

$sel:schedule:CreateCrawler', createCrawler_schedule - A cron expression used to specify the schedule (see Time-Based Schedules for Jobs and Crawlers. For example, to run something every day at 12:15 UTC, you would specify: cron(15 12 * * ? *).

$sel:crawlerSecurityConfiguration:CreateCrawler', createCrawler_crawlerSecurityConfiguration - The name of the SecurityConfiguration structure to be used by this crawler.

$sel:databaseName:CreateCrawler', createCrawler_databaseName - The Glue database where results are written, such as: arn:aws:daylight:us-east-1::database/sometable/*.

$sel:name:CreateCrawler', createCrawler_name - Name of the new crawler.

$sel:role':CreateCrawler', createCrawler_role - The IAM role or Amazon Resource Name (ARN) of an IAM role used by the new crawler to access customer resources.

$sel:targets:CreateCrawler', createCrawler_targets - A list of collection of targets to crawl.

Request Lenses

createCrawler_schemaChangePolicy :: Lens' CreateCrawler (Maybe SchemaChangePolicy)

The policy for the crawler's update and deletion behavior.

createCrawler_recrawlPolicy :: Lens' CreateCrawler (Maybe RecrawlPolicy)

A policy that specifies whether to crawl the entire dataset again, or to crawl only folders that were added since the last crawler run.

createCrawler_classifiers :: Lens' CreateCrawler (Maybe [Text])

A list of custom classifiers that the user has registered. By default, all built-in classifiers are included in a crawl, but these custom classifiers always override the default classifiers for a given classification.

createCrawler_configuration :: Lens' CreateCrawler (Maybe Text)

Crawler configuration information. This versioned JSON string allows users to specify aspects of a crawler's behavior. For more information, see Configuring a Crawler.

createCrawler_lineageConfiguration :: Lens' CreateCrawler (Maybe LineageConfiguration)

Specifies data lineage configuration settings for the crawler.

createCrawler_tags :: Lens' CreateCrawler (Maybe (HashMap Text Text))

The tags to use with this crawler request. You may use tags to limit access to the crawler. For more information about tags in Glue, see Amazon Web Services Tags in Glue in the developer guide.

createCrawler_description :: Lens' CreateCrawler (Maybe Text)

A description of the new crawler.

createCrawler_tablePrefix :: Lens' CreateCrawler (Maybe Text)

The table prefix used for catalog tables that are created.

createCrawler_schedule :: Lens' CreateCrawler (Maybe Text)

A cron expression used to specify the schedule (see Time-Based Schedules for Jobs and Crawlers. For example, to run something every day at 12:15 UTC, you would specify: cron(15 12 * * ? *).

createCrawler_crawlerSecurityConfiguration :: Lens' CreateCrawler (Maybe Text)

The name of the SecurityConfiguration structure to be used by this crawler.

createCrawler_databaseName :: Lens' CreateCrawler (Maybe Text)

The Glue database where results are written, such as: arn:aws:daylight:us-east-1::database/sometable/*.

createCrawler_name :: Lens' CreateCrawler Text

Name of the new crawler.

createCrawler_role :: Lens' CreateCrawler Text

The IAM role or Amazon Resource Name (ARN) of an IAM role used by the new crawler to access customer resources.

createCrawler_targets :: Lens' CreateCrawler CrawlerTargets

A list of collection of targets to crawl.

Destructuring the Response

data CreateCrawlerResponse

See: newCreateCrawlerResponse smart constructor.

Constructors

CreateCrawlerResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateCrawlerResponse

Arguments

:: Int
$sel:httpStatus:CreateCrawlerResponse'

-> CreateCrawlerResponse

Create a value of CreateCrawlerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateCrawlerResponse', createCrawlerResponse_httpStatus - The response's http status code.

Response Lenses

createCrawlerResponse_httpStatus :: Lens' CreateCrawlerResponse Int

The response's http status code.