Network.AWS.Glue.CreateMLTransform

Description

Creates an Glue machine learning transform. This operation creates the transform and all the necessary parameters to train it.

Call this operation as the first step in the process of using a machine learning transform (such as the FindMatches transform) for deduplicating data. You can provide an optional Description, in addition to the parameters that you want to use for your algorithm.

You must also specify certain parameters for the tasks that Glue runs on your behalf as part of learning from your data and creating a high-quality machine learning transform. These parameters include Role, and optionally, AllocatedCapacity, Timeout, and MaxRetries. For more information, see Jobs.

Creating a Request

data CreateMLTransform

See: newCreateMLTransform smart constructor.

Constructors

CreateMLTransform'

Fields

transformEncryption :: Maybe TransformEncryption

The encryption-at-rest settings of the transform that apply to accessing user data. Machine learning transforms can access user data encrypted in Amazon S3 using KMS.

timeout :: Maybe Natural

The timeout of the task run for this transform in minutes. This is the maximum time that a task run for this transform can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours).

maxCapacity :: Maybe Double

The number of Glue data processing units (DPUs) that are allocated to task runs for this transform. You can allocate from 2 to 100 DPUs; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

MaxCapacity is a mutually exclusive option with NumberOfWorkers and WorkerType.

If either NumberOfWorkers or WorkerType is set, then MaxCapacity cannot be set.

If MaxCapacity is set then neither NumberOfWorkers or WorkerType can be set.

If WorkerType is set, then NumberOfWorkers is required (and vice versa).

MaxCapacity and NumberOfWorkers must both be at least 1.

When the WorkerType field is set to a value other than Standard, the MaxCapacity field is set automatically and becomes read-only.

When the WorkerType field is set to a value other than Standard, the MaxCapacity field is set automatically and becomes read-only.

numberOfWorkers :: Maybe Int

The number of workers of a defined workerType that are allocated when this task runs.

If WorkerType is set, then NumberOfWorkers is required (and vice versa).

glueVersion :: Maybe Text

This value determines which version of Glue this machine learning transform is compatible with. Glue 1.0 is recommended for most customers. If the value is not set, the Glue compatibility defaults to Glue 0.9. For more information, see Glue Versions in the developer guide.

tags :: Maybe (HashMap Text Text)

The tags to use with this machine learning transform. You may use tags to limit access to the machine learning transform. For more information about tags in Glue, see Amazon Web Services Tags in Glue in the developer guide.

workerType :: Maybe WorkerType

The type of predefined worker that is allocated when this task runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker provides 4 vCPU, 16 GB of memory and a 64GB disk, and 1 executor per worker.

For the G.2X worker type, each worker provides 8 vCPU, 32 GB of memory and a 128GB disk, and 1 executor per worker.

MaxCapacity is a mutually exclusive option with NumberOfWorkers and WorkerType.

If either NumberOfWorkers or WorkerType is set, then MaxCapacity cannot be set.

If MaxCapacity is set then neither NumberOfWorkers or WorkerType can be set.

If WorkerType is set, then NumberOfWorkers is required (and vice versa).

MaxCapacity and NumberOfWorkers must both be at least 1.

description :: Maybe Text

A description of the machine learning transform that is being defined. The default is an empty string.

maxRetries :: Maybe Int

The maximum number of times to retry a task for this transform after a task run fails.

name :: Text

The unique name that you give the transform when you create it.

inputRecordTables :: [GlueTable]

A list of Glue table definitions used by the transform.

parameters :: TransformParameters

The algorithmic parameters that are specific to the transform type used. Conditionally dependent on the transform type.

role' :: Text

The name or Amazon Resource Name (ARN) of the IAM role with the required permissions. The required permissions include both Glue service role permissions to Glue resources, and Amazon S3 permissions required by the transform.

This role needs Glue service role permissions to allow access to resources in Glue. See Attach a Policy to IAM Users That Access Glue.

This role needs permission to your Amazon Simple Storage Service (Amazon S3) sources, targets, temporary directory, scripts, and any libraries used by the task run for this transform.

newCreateMLTransform

Arguments

:: Text
$sel:name:CreateMLTransform'

-> TransformParameters
$sel:parameters:CreateMLTransform'

-> Text
$sel:role':CreateMLTransform'

-> CreateMLTransform

Create a value of CreateMLTransform with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transformEncryption:CreateMLTransform', createMLTransform_transformEncryption - The encryption-at-rest settings of the transform that apply to accessing user data. Machine learning transforms can access user data encrypted in Amazon S3 using KMS.

$sel:timeout:CreateMLTransform', createMLTransform_timeout - The timeout of the task run for this transform in minutes. This is the maximum time that a task run for this transform can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours).

$sel:maxCapacity:CreateMLTransform', createMLTransform_maxCapacity - The number of Glue data processing units (DPUs) that are allocated to task runs for this transform. You can allocate from 2 to 100 DPUs; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

MaxCapacity is a mutually exclusive option with NumberOfWorkers and WorkerType.

If either NumberOfWorkers or WorkerType is set, then MaxCapacity cannot be set.

If MaxCapacity is set then neither NumberOfWorkers or WorkerType can be set.

If WorkerType is set, then NumberOfWorkers is required (and vice versa).

MaxCapacity and NumberOfWorkers must both be at least 1.

When the WorkerType field is set to a value other than Standard, the MaxCapacity field is set automatically and becomes read-only.

When the WorkerType field is set to a value other than Standard, the MaxCapacity field is set automatically and becomes read-only.

$sel:numberOfWorkers:CreateMLTransform', createMLTransform_numberOfWorkers - The number of workers of a defined workerType that are allocated when this task runs.

If WorkerType is set, then NumberOfWorkers is required (and vice versa).

$sel:glueVersion:CreateMLTransform', createMLTransform_glueVersion - This value determines which version of Glue this machine learning transform is compatible with. Glue 1.0 is recommended for most customers. If the value is not set, the Glue compatibility defaults to Glue 0.9. For more information, see Glue Versions in the developer guide.

$sel:tags:CreateMLTransform', createMLTransform_tags - The tags to use with this machine learning transform. You may use tags to limit access to the machine learning transform. For more information about tags in Glue, see Amazon Web Services Tags in Glue in the developer guide.

$sel:workerType:CreateMLTransform', createMLTransform_workerType - The type of predefined worker that is allocated when this task runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker provides 4 vCPU, 16 GB of memory and a 64GB disk, and 1 executor per worker.

For the G.2X worker type, each worker provides 8 vCPU, 32 GB of memory and a 128GB disk, and 1 executor per worker.

MaxCapacity is a mutually exclusive option with NumberOfWorkers and WorkerType.

If either NumberOfWorkers or WorkerType is set, then MaxCapacity cannot be set.

If MaxCapacity is set then neither NumberOfWorkers or WorkerType can be set.

If WorkerType is set, then NumberOfWorkers is required (and vice versa).

MaxCapacity and NumberOfWorkers must both be at least 1.

$sel:description:CreateMLTransform', createMLTransform_description - A description of the machine learning transform that is being defined. The default is an empty string.

$sel:maxRetries:CreateMLTransform', createMLTransform_maxRetries - The maximum number of times to retry a task for this transform after a task run fails.

$sel:name:CreateMLTransform', createMLTransform_name - The unique name that you give the transform when you create it.

$sel:inputRecordTables:CreateMLTransform', createMLTransform_inputRecordTables - A list of Glue table definitions used by the transform.

$sel:parameters:CreateMLTransform', createMLTransform_parameters - The algorithmic parameters that are specific to the transform type used. Conditionally dependent on the transform type.

$sel:role':CreateMLTransform', createMLTransform_role - The name or Amazon Resource Name (ARN) of the IAM role with the required permissions. The required permissions include both Glue service role permissions to Glue resources, and Amazon S3 permissions required by the transform.

This role needs Glue service role permissions to allow access to resources in Glue. See Attach a Policy to IAM Users That Access Glue.

This role needs permission to your Amazon Simple Storage Service (Amazon S3) sources, targets, temporary directory, scripts, and any libraries used by the task run for this transform.

Request Lenses

createMLTransform_transformEncryption :: Lens' CreateMLTransform (Maybe TransformEncryption)

The encryption-at-rest settings of the transform that apply to accessing user data. Machine learning transforms can access user data encrypted in Amazon S3 using KMS.

createMLTransform_timeout :: Lens' CreateMLTransform (Maybe Natural)

The timeout of the task run for this transform in minutes. This is the maximum time that a task run for this transform can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours).

createMLTransform_maxCapacity :: Lens' CreateMLTransform (Maybe Double)

The number of Glue data processing units (DPUs) that are allocated to task runs for this transform. You can allocate from 2 to 100 DPUs; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

MaxCapacity is a mutually exclusive option with NumberOfWorkers and WorkerType.

If either NumberOfWorkers or WorkerType is set, then MaxCapacity cannot be set.

If MaxCapacity is set then neither NumberOfWorkers or WorkerType can be set.

If WorkerType is set, then NumberOfWorkers is required (and vice versa).

MaxCapacity and NumberOfWorkers must both be at least 1.

When the WorkerType field is set to a value other than Standard, the MaxCapacity field is set automatically and becomes read-only.

When the WorkerType field is set to a value other than Standard, the MaxCapacity field is set automatically and becomes read-only.

createMLTransform_numberOfWorkers :: Lens' CreateMLTransform (Maybe Int)

The number of workers of a defined workerType that are allocated when this task runs.

If WorkerType is set, then NumberOfWorkers is required (and vice versa).

createMLTransform_glueVersion :: Lens' CreateMLTransform (Maybe Text)

This value determines which version of Glue this machine learning transform is compatible with. Glue 1.0 is recommended for most customers. If the value is not set, the Glue compatibility defaults to Glue 0.9. For more information, see Glue Versions in the developer guide.

createMLTransform_tags :: Lens' CreateMLTransform (Maybe (HashMap Text Text))

The tags to use with this machine learning transform. You may use tags to limit access to the machine learning transform. For more information about tags in Glue, see Amazon Web Services Tags in Glue in the developer guide.

createMLTransform_workerType :: Lens' CreateMLTransform (Maybe WorkerType)

The type of predefined worker that is allocated when this task runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker provides 4 vCPU, 16 GB of memory and a 64GB disk, and 1 executor per worker.

For the G.2X worker type, each worker provides 8 vCPU, 32 GB of memory and a 128GB disk, and 1 executor per worker.

MaxCapacity is a mutually exclusive option with NumberOfWorkers and WorkerType.

If either NumberOfWorkers or WorkerType is set, then MaxCapacity cannot be set.

If MaxCapacity is set then neither NumberOfWorkers or WorkerType can be set.

If WorkerType is set, then NumberOfWorkers is required (and vice versa).

MaxCapacity and NumberOfWorkers must both be at least 1.

createMLTransform_description :: Lens' CreateMLTransform (Maybe Text)

A description of the machine learning transform that is being defined. The default is an empty string.

createMLTransform_maxRetries :: Lens' CreateMLTransform (Maybe Int)

The maximum number of times to retry a task for this transform after a task run fails.

createMLTransform_name :: Lens' CreateMLTransform Text

The unique name that you give the transform when you create it.

createMLTransform_inputRecordTables :: Lens' CreateMLTransform [GlueTable]

A list of Glue table definitions used by the transform.

createMLTransform_parameters :: Lens' CreateMLTransform TransformParameters

The algorithmic parameters that are specific to the transform type used. Conditionally dependent on the transform type.

createMLTransform_role :: Lens' CreateMLTransform Text

The name or Amazon Resource Name (ARN) of the IAM role with the required permissions. The required permissions include both Glue service role permissions to Glue resources, and Amazon S3 permissions required by the transform.

This role needs Glue service role permissions to allow access to resources in Glue. See Attach a Policy to IAM Users That Access Glue.

This role needs permission to your Amazon Simple Storage Service (Amazon S3) sources, targets, temporary directory, scripts, and any libraries used by the task run for this transform.

Destructuring the Response

data CreateMLTransformResponse

See: newCreateMLTransformResponse smart constructor.

Constructors

CreateMLTransformResponse'

Fields

transformId :: Maybe Text

A unique identifier that is generated for the transform.

httpStatus :: Int

The response's http status code.

newCreateMLTransformResponse

Arguments

:: Int
$sel:httpStatus:CreateMLTransformResponse'

-> CreateMLTransformResponse

Create a value of CreateMLTransformResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transformId:CreateMLTransformResponse', createMLTransformResponse_transformId - A unique identifier that is generated for the transform.

$sel:httpStatus:CreateMLTransformResponse', createMLTransformResponse_httpStatus - The response's http status code.

Response Lenses

createMLTransformResponse_transformId :: Lens' CreateMLTransformResponse (Maybe Text)

A unique identifier that is generated for the transform.

createMLTransformResponse_httpStatus :: Lens' CreateMLTransformResponse Int

The response's http status code.