Network.AWS.Glue.CreatePartitionIndex

Description

Creates a specified partition index in an existing table.

Creating a Request

data CreatePartitionIndex

See: newCreatePartitionIndex smart constructor.

Constructors

CreatePartitionIndex'

Fields

catalogId :: Maybe Text

The catalog ID where the table resides.

databaseName :: Text

Specifies the name of a database in which you want to create a partition index.

tableName :: Text

Specifies the name of a table in which you want to create a partition index.

partitionIndex :: PartitionIndex

Specifies a PartitionIndex structure to create a partition index in an existing table.

newCreatePartitionIndex

Arguments

:: Text
$sel:databaseName:CreatePartitionIndex'

-> Text
$sel:tableName:CreatePartitionIndex'

-> PartitionIndex
$sel:partitionIndex:CreatePartitionIndex'

-> CreatePartitionIndex

Create a value of CreatePartitionIndex with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:CreatePartitionIndex', createPartitionIndex_catalogId - The catalog ID where the table resides.

$sel:databaseName:CreatePartitionIndex', createPartitionIndex_databaseName - Specifies the name of a database in which you want to create a partition index.

$sel:tableName:CreatePartitionIndex', createPartitionIndex_tableName - Specifies the name of a table in which you want to create a partition index.

$sel:partitionIndex:CreatePartitionIndex', createPartitionIndex_partitionIndex - Specifies a PartitionIndex structure to create a partition index in an existing table.

Request Lenses

createPartitionIndex_catalogId :: Lens' CreatePartitionIndex (Maybe Text)

The catalog ID where the table resides.

createPartitionIndex_databaseName :: Lens' CreatePartitionIndex Text

Specifies the name of a database in which you want to create a partition index.

createPartitionIndex_tableName :: Lens' CreatePartitionIndex Text

Specifies the name of a table in which you want to create a partition index.

createPartitionIndex_partitionIndex :: Lens' CreatePartitionIndex PartitionIndex

Specifies a PartitionIndex structure to create a partition index in an existing table.

Destructuring the Response

data CreatePartitionIndexResponse

See: newCreatePartitionIndexResponse smart constructor.

Constructors

CreatePartitionIndexResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreatePartitionIndexResponse

Arguments

:: Int
$sel:httpStatus:CreatePartitionIndexResponse'

-> CreatePartitionIndexResponse

Create a value of CreatePartitionIndexResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreatePartitionIndexResponse', createPartitionIndexResponse_httpStatus - The response's http status code.

Response Lenses

createPartitionIndexResponse_httpStatus :: Lens' CreatePartitionIndexResponse Int

The response's http status code.