Network.AWS.Glue.CreateRegistry

Description

Creates a new registry which may be used to hold a collection of schemas.

Creating a Request

data CreateRegistry

See: newCreateRegistry smart constructor.

Constructors

CreateRegistry'

Fields

tags :: Maybe (HashMap Text Text)

Amazon Web Services tags that contain a key value pair and may be searched by console, command line, or API.

description :: Maybe Text

A description of the registry. If description is not provided, there will not be any default value for this.

registryName :: Text

Name of the registry to be created of max length of 255, and may only contain letters, numbers, hyphen, underscore, dollar sign, or hash mark. No whitespace.

newCreateRegistry

Arguments

:: Text
$sel:registryName:CreateRegistry'

-> CreateRegistry

Create a value of CreateRegistry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateRegistry', createRegistry_tags - Amazon Web Services tags that contain a key value pair and may be searched by console, command line, or API.

$sel:description:CreateRegistry', createRegistry_description - A description of the registry. If description is not provided, there will not be any default value for this.

$sel:registryName:CreateRegistry', createRegistry_registryName - Name of the registry to be created of max length of 255, and may only contain letters, numbers, hyphen, underscore, dollar sign, or hash mark. No whitespace.

Request Lenses

createRegistry_tags :: Lens' CreateRegistry (Maybe (HashMap Text Text))

Amazon Web Services tags that contain a key value pair and may be searched by console, command line, or API.

createRegistry_description :: Lens' CreateRegistry (Maybe Text)

A description of the registry. If description is not provided, there will not be any default value for this.

createRegistry_registryName :: Lens' CreateRegistry Text

Name of the registry to be created of max length of 255, and may only contain letters, numbers, hyphen, underscore, dollar sign, or hash mark. No whitespace.

Destructuring the Response

data CreateRegistryResponse

See: newCreateRegistryResponse smart constructor.

Constructors

CreateRegistryResponse'

Fields

registryName :: Maybe Text

The name of the registry.

tags :: Maybe (HashMap Text Text)

The tags for the registry.

description :: Maybe Text

A description of the registry.

registryArn :: Maybe Text

The Amazon Resource Name (ARN) of the newly created registry.

httpStatus :: Int

The response's http status code.

newCreateRegistryResponse

Arguments

:: Int
$sel:httpStatus:CreateRegistryResponse'

-> CreateRegistryResponse

Create a value of CreateRegistryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryName:CreateRegistry', createRegistryResponse_registryName - The name of the registry.

$sel:tags:CreateRegistry', createRegistryResponse_tags - The tags for the registry.

$sel:description:CreateRegistry', createRegistryResponse_description - A description of the registry.

$sel:registryArn:CreateRegistryResponse', createRegistryResponse_registryArn - The Amazon Resource Name (ARN) of the newly created registry.

$sel:httpStatus:CreateRegistryResponse', createRegistryResponse_httpStatus - The response's http status code.

Response Lenses

createRegistryResponse_registryName :: Lens' CreateRegistryResponse (Maybe Text)

The name of the registry.

createRegistryResponse_tags :: Lens' CreateRegistryResponse (Maybe (HashMap Text Text))

The tags for the registry.

createRegistryResponse_description :: Lens' CreateRegistryResponse (Maybe Text)

A description of the registry.

createRegistryResponse_registryArn :: Lens' CreateRegistryResponse (Maybe Text)

The Amazon Resource Name (ARN) of the newly created registry.

createRegistryResponse_httpStatus :: Lens' CreateRegistryResponse Int

The response's http status code.