Network.AWS.Glue.CreateSchema

Description

Creates a new schema set and registers the schema definition. Returns an error if the schema set already exists without actually registering the version.

When the schema set is created, a version checkpoint will be set to the first version. Compatibility mode "DISABLED" restricts any additional schema versions from being added after the first schema version. For all other compatibility modes, validation of compatibility settings will be applied only from the second version onwards when the RegisterSchemaVersion API is used.

When this API is called without a RegistryId, this will create an entry for a "default-registry" in the registry database tables, if it is not already present.

Creating a Request

data CreateSchema

See: newCreateSchema smart constructor.

Constructors

CreateSchema'

Fields

schemaDefinition :: Maybe Text

The schema definition using the DataFormat setting for SchemaName.

registryId :: Maybe RegistryId

This is a wrapper shape to contain the registry identity fields. If this is not provided, the default registry will be used. The ARN format for the same will be: arn:aws:glue:us-east-2:<customer id>:registry/default-registry:random-5-letter-id.

tags :: Maybe (HashMap Text Text)

Amazon Web Services tags that contain a key value pair and may be searched by console, command line, or API. If specified, follows the Amazon Web Services tags-on-create pattern.

compatibility :: Maybe Compatibility

The compatibility mode of the schema. The possible values are:

NONE: No compatibility mode applies. You can use this choice in development scenarios or if you do not know the compatibility mode that you want to apply to schemas. Any new version added will be accepted without undergoing a compatibility check.

DISABLED: This compatibility choice prevents versioning for a particular schema. You can use this choice to prevent future versioning of a schema.

BACKWARD: This compatibility choice is recommended as it allows data receivers to read both the current and one previous schema version. This means that for instance, a new schema version cannot drop data fields or change the type of these fields, so they can't be read by readers using the previous version.

BACKWARD_ALL: This compatibility choice allows data receivers to read both the current and all previous schema versions. You can use this choice when you need to delete fields or add optional fields, and check compatibility against all previous schema versions.

FORWARD: This compatibility choice allows data receivers to read both the current and one next schema version, but not necessarily later versions. You can use this choice when you need to add fields or delete optional fields, but only check compatibility against the last schema version.

FORWARD_ALL: This compatibility choice allows data receivers to read written by producers of any new registered schema. You can use this choice when you need to add fields or delete optional fields, and check compatibility against all previous schema versions.

FULL: This compatibility choice allows data receivers to read data written by producers using the previous or next version of the schema, but not necessarily earlier or later versions. You can use this choice when you need to add or remove optional fields, but only check compatibility against the last schema version.

FULL_ALL: This compatibility choice allows data receivers to read data written by producers using all previous schema versions. You can use this choice when you need to add or remove optional fields, and check compatibility against all previous schema versions.

description :: Maybe Text

An optional description of the schema. If description is not provided, there will not be any automatic default value for this.

schemaName :: Text

Name of the schema to be created of max length of 255, and may only contain letters, numbers, hyphen, underscore, dollar sign, or hash mark. No whitespace.

dataFormat :: DataFormat

The data format of the schema definition. Currently AVRO and JSON are supported.

newCreateSchema

Arguments

:: Text
$sel:schemaName:CreateSchema'

-> DataFormat
$sel:dataFormat:CreateSchema'

-> CreateSchema

Create a value of CreateSchema with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaDefinition:CreateSchema', createSchema_schemaDefinition - The schema definition using the DataFormat setting for SchemaName.

$sel:registryId:CreateSchema', createSchema_registryId - This is a wrapper shape to contain the registry identity fields. If this is not provided, the default registry will be used. The ARN format for the same will be: arn:aws:glue:us-east-2:<customer id>:registry/default-registry:random-5-letter-id.

$sel:tags:CreateSchema', createSchema_tags - Amazon Web Services tags that contain a key value pair and may be searched by console, command line, or API. If specified, follows the Amazon Web Services tags-on-create pattern.

$sel:compatibility:CreateSchema', createSchema_compatibility - The compatibility mode of the schema. The possible values are:

NONE: No compatibility mode applies. You can use this choice in development scenarios or if you do not know the compatibility mode that you want to apply to schemas. Any new version added will be accepted without undergoing a compatibility check.

DISABLED: This compatibility choice prevents versioning for a particular schema. You can use this choice to prevent future versioning of a schema.

BACKWARD: This compatibility choice is recommended as it allows data receivers to read both the current and one previous schema version. This means that for instance, a new schema version cannot drop data fields or change the type of these fields, so they can't be read by readers using the previous version.

BACKWARD_ALL: This compatibility choice allows data receivers to read both the current and all previous schema versions. You can use this choice when you need to delete fields or add optional fields, and check compatibility against all previous schema versions.

FORWARD: This compatibility choice allows data receivers to read both the current and one next schema version, but not necessarily later versions. You can use this choice when you need to add fields or delete optional fields, but only check compatibility against the last schema version.

FORWARD_ALL: This compatibility choice allows data receivers to read written by producers of any new registered schema. You can use this choice when you need to add fields or delete optional fields, and check compatibility against all previous schema versions.

FULL: This compatibility choice allows data receivers to read data written by producers using the previous or next version of the schema, but not necessarily earlier or later versions. You can use this choice when you need to add or remove optional fields, but only check compatibility against the last schema version.

FULL_ALL: This compatibility choice allows data receivers to read data written by producers using all previous schema versions. You can use this choice when you need to add or remove optional fields, and check compatibility against all previous schema versions.

$sel:description:CreateSchema', createSchema_description - An optional description of the schema. If description is not provided, there will not be any automatic default value for this.

$sel:schemaName:CreateSchema', createSchema_schemaName - Name of the schema to be created of max length of 255, and may only contain letters, numbers, hyphen, underscore, dollar sign, or hash mark. No whitespace.

$sel:dataFormat:CreateSchema', createSchema_dataFormat - The data format of the schema definition. Currently AVRO and JSON are supported.

Request Lenses

createSchema_schemaDefinition :: Lens' CreateSchema (Maybe Text)

The schema definition using the DataFormat setting for SchemaName.

createSchema_registryId :: Lens' CreateSchema (Maybe RegistryId)

This is a wrapper shape to contain the registry identity fields. If this is not provided, the default registry will be used. The ARN format for the same will be: arn:aws:glue:us-east-2:<customer id>:registry/default-registry:random-5-letter-id.

createSchema_tags :: Lens' CreateSchema (Maybe (HashMap Text Text))

Amazon Web Services tags that contain a key value pair and may be searched by console, command line, or API. If specified, follows the Amazon Web Services tags-on-create pattern.

createSchema_compatibility :: Lens' CreateSchema (Maybe Compatibility)

The compatibility mode of the schema. The possible values are:

NONE: No compatibility mode applies. You can use this choice in development scenarios or if you do not know the compatibility mode that you want to apply to schemas. Any new version added will be accepted without undergoing a compatibility check.

DISABLED: This compatibility choice prevents versioning for a particular schema. You can use this choice to prevent future versioning of a schema.

BACKWARD: This compatibility choice is recommended as it allows data receivers to read both the current and one previous schema version. This means that for instance, a new schema version cannot drop data fields or change the type of these fields, so they can't be read by readers using the previous version.

BACKWARD_ALL: This compatibility choice allows data receivers to read both the current and all previous schema versions. You can use this choice when you need to delete fields or add optional fields, and check compatibility against all previous schema versions.

FORWARD: This compatibility choice allows data receivers to read both the current and one next schema version, but not necessarily later versions. You can use this choice when you need to add fields or delete optional fields, but only check compatibility against the last schema version.

FORWARD_ALL: This compatibility choice allows data receivers to read written by producers of any new registered schema. You can use this choice when you need to add fields or delete optional fields, and check compatibility against all previous schema versions.

FULL: This compatibility choice allows data receivers to read data written by producers using the previous or next version of the schema, but not necessarily earlier or later versions. You can use this choice when you need to add or remove optional fields, but only check compatibility against the last schema version.

FULL_ALL: This compatibility choice allows data receivers to read data written by producers using all previous schema versions. You can use this choice when you need to add or remove optional fields, and check compatibility against all previous schema versions.

createSchema_description :: Lens' CreateSchema (Maybe Text)

An optional description of the schema. If description is not provided, there will not be any automatic default value for this.

createSchema_schemaName :: Lens' CreateSchema Text

Name of the schema to be created of max length of 255, and may only contain letters, numbers, hyphen, underscore, dollar sign, or hash mark. No whitespace.

createSchema_dataFormat :: Lens' CreateSchema DataFormat

The data format of the schema definition. Currently AVRO and JSON are supported.

Destructuring the Response

data CreateSchemaResponse

See: newCreateSchemaResponse smart constructor.

Constructors

CreateSchemaResponse'

Fields

schemaArn :: Maybe Text

The Amazon Resource Name (ARN) of the schema.

nextSchemaVersion :: Maybe Natural

The next version of the schema associated with the returned schema definition.

schemaVersionId :: Maybe Text

The unique identifier of the first schema version.

schemaCheckpoint :: Maybe Natural

The version number of the checkpoint (the last time the compatibility mode was changed).

dataFormat :: Maybe DataFormat

The data format of the schema definition. Currently AVRO and JSON are supported.

registryName :: Maybe Text

The name of the registry.

schemaVersionStatus :: Maybe SchemaVersionStatus

The status of the first schema version created.

tags :: Maybe (HashMap Text Text)

The tags for the schema.

compatibility :: Maybe Compatibility

The schema compatibility mode.

description :: Maybe Text

A description of the schema if specified when created.

schemaName :: Maybe Text

The name of the schema.

registryArn :: Maybe Text

The Amazon Resource Name (ARN) of the registry.

latestSchemaVersion :: Maybe Natural

The latest version of the schema associated with the returned schema definition.

schemaStatus :: Maybe SchemaStatus

The status of the schema.

httpStatus :: Int

The response's http status code.

newCreateSchemaResponse

Arguments

:: Int
$sel:httpStatus:CreateSchemaResponse'

-> CreateSchemaResponse

Create a value of CreateSchemaResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaArn:CreateSchemaResponse', createSchemaResponse_schemaArn - The Amazon Resource Name (ARN) of the schema.

$sel:nextSchemaVersion:CreateSchemaResponse', createSchemaResponse_nextSchemaVersion - The next version of the schema associated with the returned schema definition.

$sel:schemaVersionId:CreateSchemaResponse', createSchemaResponse_schemaVersionId - The unique identifier of the first schema version.

$sel:schemaCheckpoint:CreateSchemaResponse', createSchemaResponse_schemaCheckpoint - The version number of the checkpoint (the last time the compatibility mode was changed).

$sel:dataFormat:CreateSchema', createSchemaResponse_dataFormat - The data format of the schema definition. Currently AVRO and JSON are supported.

$sel:registryName:CreateSchemaResponse', createSchemaResponse_registryName - The name of the registry.

$sel:schemaVersionStatus:CreateSchemaResponse', createSchemaResponse_schemaVersionStatus - The status of the first schema version created.

$sel:tags:CreateSchema', createSchemaResponse_tags - The tags for the schema.

$sel:compatibility:CreateSchema', createSchemaResponse_compatibility - The schema compatibility mode.

$sel:description:CreateSchema', createSchemaResponse_description - A description of the schema if specified when created.

$sel:schemaName:CreateSchema', createSchemaResponse_schemaName - The name of the schema.

$sel:registryArn:CreateSchemaResponse', createSchemaResponse_registryArn - The Amazon Resource Name (ARN) of the registry.

$sel:latestSchemaVersion:CreateSchemaResponse', createSchemaResponse_latestSchemaVersion - The latest version of the schema associated with the returned schema definition.

$sel:schemaStatus:CreateSchemaResponse', createSchemaResponse_schemaStatus - The status of the schema.

$sel:httpStatus:CreateSchemaResponse', createSchemaResponse_httpStatus - The response's http status code.

Response Lenses

createSchemaResponse_schemaArn :: Lens' CreateSchemaResponse (Maybe Text)

The Amazon Resource Name (ARN) of the schema.

createSchemaResponse_nextSchemaVersion :: Lens' CreateSchemaResponse (Maybe Natural)

The next version of the schema associated with the returned schema definition.

createSchemaResponse_schemaVersionId :: Lens' CreateSchemaResponse (Maybe Text)

The unique identifier of the first schema version.

createSchemaResponse_schemaCheckpoint :: Lens' CreateSchemaResponse (Maybe Natural)

The version number of the checkpoint (the last time the compatibility mode was changed).

createSchemaResponse_dataFormat :: Lens' CreateSchemaResponse (Maybe DataFormat)

The data format of the schema definition. Currently AVRO and JSON are supported.

createSchemaResponse_registryName :: Lens' CreateSchemaResponse (Maybe Text)

The name of the registry.

createSchemaResponse_schemaVersionStatus :: Lens' CreateSchemaResponse (Maybe SchemaVersionStatus)

The status of the first schema version created.

createSchemaResponse_tags :: Lens' CreateSchemaResponse (Maybe (HashMap Text Text))

The tags for the schema.

createSchemaResponse_compatibility :: Lens' CreateSchemaResponse (Maybe Compatibility)

The schema compatibility mode.

createSchemaResponse_description :: Lens' CreateSchemaResponse (Maybe Text)

A description of the schema if specified when created.

createSchemaResponse_schemaName :: Lens' CreateSchemaResponse (Maybe Text)

The name of the schema.

createSchemaResponse_registryArn :: Lens' CreateSchemaResponse (Maybe Text)

The Amazon Resource Name (ARN) of the registry.

createSchemaResponse_latestSchemaVersion :: Lens' CreateSchemaResponse (Maybe Natural)

The latest version of the schema associated with the returned schema definition.

createSchemaResponse_schemaStatus :: Lens' CreateSchemaResponse (Maybe SchemaStatus)

The status of the schema.

createSchemaResponse_httpStatus :: Lens' CreateSchemaResponse Int

The response's http status code.