Network.AWS.Glue.CreateScript

Description

Transforms a directed acyclic graph (DAG) into code.

Creating a Request

data CreateScript

See: newCreateScript smart constructor.

Constructors

CreateScript'

Fields

dagNodes :: Maybe [CodeGenNode]

A list of the nodes in the DAG.

language :: Maybe Language

The programming language of the resulting code from the DAG.

dagEdges :: Maybe [CodeGenEdge]

A list of the edges in the DAG.

newCreateScript :: CreateScript

Create a value of CreateScript with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dagNodes:CreateScript', createScript_dagNodes - A list of the nodes in the DAG.

$sel:language:CreateScript', createScript_language - The programming language of the resulting code from the DAG.

$sel:dagEdges:CreateScript', createScript_dagEdges - A list of the edges in the DAG.

Request Lenses

createScript_dagNodes :: Lens' CreateScript (Maybe [CodeGenNode])

A list of the nodes in the DAG.

createScript_language :: Lens' CreateScript (Maybe Language)

The programming language of the resulting code from the DAG.

createScript_dagEdges :: Lens' CreateScript (Maybe [CodeGenEdge])

A list of the edges in the DAG.

Destructuring the Response

data CreateScriptResponse

See: newCreateScriptResponse smart constructor.

Constructors

CreateScriptResponse'

Fields

pythonScript :: Maybe Text

The Python script generated from the DAG.

scalaCode :: Maybe Text

The Scala code generated from the DAG.

httpStatus :: Int

The response's http status code.

newCreateScriptResponse

Arguments

:: Int
$sel:httpStatus:CreateScriptResponse'

-> CreateScriptResponse

Create a value of CreateScriptResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pythonScript:CreateScriptResponse', createScriptResponse_pythonScript - The Python script generated from the DAG.

$sel:scalaCode:CreateScriptResponse', createScriptResponse_scalaCode - The Scala code generated from the DAG.

$sel:httpStatus:CreateScriptResponse', createScriptResponse_httpStatus - The response's http status code.

Response Lenses

createScriptResponse_pythonScript :: Lens' CreateScriptResponse (Maybe Text)

The Python script generated from the DAG.

createScriptResponse_scalaCode :: Lens' CreateScriptResponse (Maybe Text)

The Scala code generated from the DAG.

createScriptResponse_httpStatus :: Lens' CreateScriptResponse Int

The response's http status code.