Network.AWS.Glue.CreateTable

Description

Creates a new table definition in the Data Catalog.

Creating a Request

data CreateTable

See: newCreateTable smart constructor.

Constructors

CreateTable'

Fields

catalogId :: Maybe Text

The ID of the Data Catalog in which to create the Table. If none is supplied, the Amazon Web Services account ID is used by default.

partitionIndexes :: Maybe [PartitionIndex]

A list of partition indexes, PartitionIndex structures, to create in the table.

databaseName :: Text

The catalog database in which to create the new table. For Hive compatibility, this name is entirely lowercase.

tableInput :: TableInput

The TableInput object that defines the metadata table to create in the catalog.

newCreateTable

Arguments

:: Text
$sel:databaseName:CreateTable'

-> TableInput
$sel:tableInput:CreateTable'

-> CreateTable

Create a value of CreateTable with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:CreateTable', createTable_catalogId - The ID of the Data Catalog in which to create the Table. If none is supplied, the Amazon Web Services account ID is used by default.

$sel:partitionIndexes:CreateTable', createTable_partitionIndexes - A list of partition indexes, PartitionIndex structures, to create in the table.

$sel:databaseName:CreateTable', createTable_databaseName - The catalog database in which to create the new table. For Hive compatibility, this name is entirely lowercase.

$sel:tableInput:CreateTable', createTable_tableInput - The TableInput object that defines the metadata table to create in the catalog.

Request Lenses

createTable_catalogId :: Lens' CreateTable (Maybe Text)

The ID of the Data Catalog in which to create the Table. If none is supplied, the Amazon Web Services account ID is used by default.

createTable_partitionIndexes :: Lens' CreateTable (Maybe [PartitionIndex])

A list of partition indexes, PartitionIndex structures, to create in the table.

createTable_databaseName :: Lens' CreateTable Text

The catalog database in which to create the new table. For Hive compatibility, this name is entirely lowercase.

createTable_tableInput :: Lens' CreateTable TableInput

The TableInput object that defines the metadata table to create in the catalog.

Destructuring the Response

data CreateTableResponse

See: newCreateTableResponse smart constructor.

Constructors

CreateTableResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateTableResponse

Arguments

:: Int
$sel:httpStatus:CreateTableResponse'

-> CreateTableResponse

Create a value of CreateTableResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateTableResponse', createTableResponse_httpStatus - The response's http status code.

Response Lenses

createTableResponse_httpStatus :: Lens' CreateTableResponse Int

The response's http status code.