Network.AWS.Glue.CreateTrigger

Description

Creates a new trigger.

Creating a Request

data CreateTrigger

See: newCreateTrigger smart constructor.

Constructors

CreateTrigger'

Fields

workflowName :: Maybe Text

The name of the workflow associated with the trigger.

startOnCreation :: Maybe Bool

Set to true to start SCHEDULED and CONDITIONAL triggers when created. True is not supported for ON_DEMAND triggers.

tags :: Maybe (HashMap Text Text)

The tags to use with this trigger. You may use tags to limit access to the trigger. For more information about tags in Glue, see Amazon Web Services Tags in Glue in the developer guide.

predicate :: Maybe Predicate

A predicate to specify when the new trigger should fire.

This field is required when the trigger type is CONDITIONAL.

description :: Maybe Text

A description of the new trigger.

eventBatchingCondition :: Maybe EventBatchingCondition

Batch condition that must be met (specified number of events received or batch time window expired) before EventBridge event trigger fires.

schedule :: Maybe Text

A cron expression used to specify the schedule (see Time-Based Schedules for Jobs and Crawlers. For example, to run something every day at 12:15 UTC, you would specify: cron(15 12 * * ? *).

This field is required when the trigger type is SCHEDULED.

name :: Text

The name of the trigger.

type' :: TriggerType

The type of the new trigger.

actions :: [Action]

The actions initiated by this trigger when it fires.

newCreateTrigger

Arguments

:: Text
$sel:name:CreateTrigger'

-> TriggerType
$sel:type':CreateTrigger'

-> CreateTrigger

Create a value of CreateTrigger with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workflowName:CreateTrigger', createTrigger_workflowName - The name of the workflow associated with the trigger.

$sel:startOnCreation:CreateTrigger', createTrigger_startOnCreation - Set to true to start SCHEDULED and CONDITIONAL triggers when created. True is not supported for ON_DEMAND triggers.

$sel:tags:CreateTrigger', createTrigger_tags - The tags to use with this trigger. You may use tags to limit access to the trigger. For more information about tags in Glue, see Amazon Web Services Tags in Glue in the developer guide.

$sel:predicate:CreateTrigger', createTrigger_predicate - A predicate to specify when the new trigger should fire.

This field is required when the trigger type is CONDITIONAL.

$sel:description:CreateTrigger', createTrigger_description - A description of the new trigger.

$sel:eventBatchingCondition:CreateTrigger', createTrigger_eventBatchingCondition - Batch condition that must be met (specified number of events received or batch time window expired) before EventBridge event trigger fires.

$sel:schedule:CreateTrigger', createTrigger_schedule - A cron expression used to specify the schedule (see Time-Based Schedules for Jobs and Crawlers. For example, to run something every day at 12:15 UTC, you would specify: cron(15 12 * * ? *).

This field is required when the trigger type is SCHEDULED.

$sel:name:CreateTrigger', createTrigger_name - The name of the trigger.

$sel:type':CreateTrigger', createTrigger_type - The type of the new trigger.

$sel:actions:CreateTrigger', createTrigger_actions - The actions initiated by this trigger when it fires.

Request Lenses

createTrigger_workflowName :: Lens' CreateTrigger (Maybe Text)

The name of the workflow associated with the trigger.

createTrigger_startOnCreation :: Lens' CreateTrigger (Maybe Bool)

Set to true to start SCHEDULED and CONDITIONAL triggers when created. True is not supported for ON_DEMAND triggers.

createTrigger_tags :: Lens' CreateTrigger (Maybe (HashMap Text Text))

The tags to use with this trigger. You may use tags to limit access to the trigger. For more information about tags in Glue, see Amazon Web Services Tags in Glue in the developer guide.

createTrigger_predicate :: Lens' CreateTrigger (Maybe Predicate)

A predicate to specify when the new trigger should fire.

This field is required when the trigger type is CONDITIONAL.

createTrigger_description :: Lens' CreateTrigger (Maybe Text)

A description of the new trigger.

createTrigger_eventBatchingCondition :: Lens' CreateTrigger (Maybe EventBatchingCondition)

Batch condition that must be met (specified number of events received or batch time window expired) before EventBridge event trigger fires.

createTrigger_schedule :: Lens' CreateTrigger (Maybe Text)

A cron expression used to specify the schedule (see Time-Based Schedules for Jobs and Crawlers. For example, to run something every day at 12:15 UTC, you would specify: cron(15 12 * * ? *).

This field is required when the trigger type is SCHEDULED.

createTrigger_name :: Lens' CreateTrigger Text

The name of the trigger.

createTrigger_type :: Lens' CreateTrigger TriggerType

The type of the new trigger.

createTrigger_actions :: Lens' CreateTrigger [Action]

The actions initiated by this trigger when it fires.

Destructuring the Response

data CreateTriggerResponse

See: newCreateTriggerResponse smart constructor.

Constructors

CreateTriggerResponse'

Fields

name :: Maybe Text

The name of the trigger.

httpStatus :: Int

The response's http status code.

newCreateTriggerResponse

Arguments

:: Int
$sel:httpStatus:CreateTriggerResponse'

-> CreateTriggerResponse

Create a value of CreateTriggerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:CreateTrigger', createTriggerResponse_name - The name of the trigger.

$sel:httpStatus:CreateTriggerResponse', createTriggerResponse_httpStatus - The response's http status code.

Response Lenses

createTriggerResponse_name :: Lens' CreateTriggerResponse (Maybe Text)

The name of the trigger.

createTriggerResponse_httpStatus :: Lens' CreateTriggerResponse Int

The response's http status code.