Network.AWS.Glue.CreateWorkflow

Description

Creates a new workflow.

Creating a Request

data CreateWorkflow

See: newCreateWorkflow smart constructor.

Constructors

CreateWorkflow'

Fields

maxConcurrentRuns :: Maybe Int

You can use this parameter to prevent unwanted multiple updates to data, to control costs, or in some cases, to prevent exceeding the maximum number of concurrent runs of any of the component jobs. If you leave this parameter blank, there is no limit to the number of concurrent workflow runs.

defaultRunProperties :: Maybe (HashMap Text Text)

A collection of properties to be used as part of each execution of the workflow.

description :: Maybe Text

A description of the workflow.

tags :: Maybe (HashMap Text Text)

The tags to be used with this workflow.

name :: Text

The name to be assigned to the workflow. It should be unique within your account.

newCreateWorkflow

Arguments

:: Text
$sel:name:CreateWorkflow'

-> CreateWorkflow

Create a value of CreateWorkflow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxConcurrentRuns:CreateWorkflow', createWorkflow_maxConcurrentRuns - You can use this parameter to prevent unwanted multiple updates to data, to control costs, or in some cases, to prevent exceeding the maximum number of concurrent runs of any of the component jobs. If you leave this parameter blank, there is no limit to the number of concurrent workflow runs.

$sel:defaultRunProperties:CreateWorkflow', createWorkflow_defaultRunProperties - A collection of properties to be used as part of each execution of the workflow.

$sel:description:CreateWorkflow', createWorkflow_description - A description of the workflow.

$sel:tags:CreateWorkflow', createWorkflow_tags - The tags to be used with this workflow.

$sel:name:CreateWorkflow', createWorkflow_name - The name to be assigned to the workflow. It should be unique within your account.

Request Lenses

createWorkflow_maxConcurrentRuns :: Lens' CreateWorkflow (Maybe Int)

You can use this parameter to prevent unwanted multiple updates to data, to control costs, or in some cases, to prevent exceeding the maximum number of concurrent runs of any of the component jobs. If you leave this parameter blank, there is no limit to the number of concurrent workflow runs.

createWorkflow_defaultRunProperties :: Lens' CreateWorkflow (Maybe (HashMap Text Text))

A collection of properties to be used as part of each execution of the workflow.

createWorkflow_description :: Lens' CreateWorkflow (Maybe Text)

A description of the workflow.

createWorkflow_tags :: Lens' CreateWorkflow (Maybe (HashMap Text Text))

The tags to be used with this workflow.

createWorkflow_name :: Lens' CreateWorkflow Text

The name to be assigned to the workflow. It should be unique within your account.

Destructuring the Response

data CreateWorkflowResponse

See: newCreateWorkflowResponse smart constructor.

Constructors

CreateWorkflowResponse'

Fields

name :: Maybe Text

The name of the workflow which was provided as part of the request.

httpStatus :: Int

The response's http status code.

newCreateWorkflowResponse

Arguments

:: Int
$sel:httpStatus:CreateWorkflowResponse'

-> CreateWorkflowResponse

Create a value of CreateWorkflowResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:CreateWorkflow', createWorkflowResponse_name - The name of the workflow which was provided as part of the request.

$sel:httpStatus:CreateWorkflowResponse', createWorkflowResponse_httpStatus - The response's http status code.

Response Lenses

createWorkflowResponse_name :: Lens' CreateWorkflowResponse (Maybe Text)

The name of the workflow which was provided as part of the request.

createWorkflowResponse_httpStatus :: Lens' CreateWorkflowResponse Int

The response's http status code.