Network.AWS.Glue.DeleteColumnStatisticsForPartition

Description

Delete the partition column statistics of a column.

The Identity and Access Management (IAM) permission required for this operation is DeletePartition.

Creating a Request

data DeleteColumnStatisticsForPartition

See: newDeleteColumnStatisticsForPartition smart constructor.

Constructors

DeleteColumnStatisticsForPartition'

Fields

catalogId :: Maybe Text

The ID of the Data Catalog where the partitions in question reside. If none is supplied, the Amazon Web Services account ID is used by default.

databaseName :: Text

The name of the catalog database where the partitions reside.

tableName :: Text

The name of the partitions' table.

partitionValues :: [Text]

A list of partition values identifying the partition.

columnName :: Text

Name of the column.

newDeleteColumnStatisticsForPartition

Arguments

:: Text
$sel:databaseName:DeleteColumnStatisticsForPartition'

-> Text
$sel:tableName:DeleteColumnStatisticsForPartition'

-> Text
$sel:columnName:DeleteColumnStatisticsForPartition'

-> DeleteColumnStatisticsForPartition

Create a value of DeleteColumnStatisticsForPartition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:DeleteColumnStatisticsForPartition', deleteColumnStatisticsForPartition_catalogId - The ID of the Data Catalog where the partitions in question reside. If none is supplied, the Amazon Web Services account ID is used by default.

$sel:databaseName:DeleteColumnStatisticsForPartition', deleteColumnStatisticsForPartition_databaseName - The name of the catalog database where the partitions reside.

$sel:tableName:DeleteColumnStatisticsForPartition', deleteColumnStatisticsForPartition_tableName - The name of the partitions' table.

$sel:partitionValues:DeleteColumnStatisticsForPartition', deleteColumnStatisticsForPartition_partitionValues - A list of partition values identifying the partition.

$sel:columnName:DeleteColumnStatisticsForPartition', deleteColumnStatisticsForPartition_columnName - Name of the column.

Request Lenses

deleteColumnStatisticsForPartition_catalogId :: Lens' DeleteColumnStatisticsForPartition (Maybe Text)

The ID of the Data Catalog where the partitions in question reside. If none is supplied, the Amazon Web Services account ID is used by default.

deleteColumnStatisticsForPartition_databaseName :: Lens' DeleteColumnStatisticsForPartition Text

The name of the catalog database where the partitions reside.

deleteColumnStatisticsForPartition_tableName :: Lens' DeleteColumnStatisticsForPartition Text

The name of the partitions' table.

deleteColumnStatisticsForPartition_partitionValues :: Lens' DeleteColumnStatisticsForPartition [Text]

A list of partition values identifying the partition.

deleteColumnStatisticsForPartition_columnName :: Lens' DeleteColumnStatisticsForPartition Text

Name of the column.

Destructuring the Response

data DeleteColumnStatisticsForPartitionResponse

See: newDeleteColumnStatisticsForPartitionResponse smart constructor.

Constructors

DeleteColumnStatisticsForPartitionResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteColumnStatisticsForPartitionResponse

Arguments

:: Int
$sel:httpStatus:DeleteColumnStatisticsForPartitionResponse'

-> DeleteColumnStatisticsForPartitionResponse

Create a value of DeleteColumnStatisticsForPartitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteColumnStatisticsForPartitionResponse', deleteColumnStatisticsForPartitionResponse_httpStatus - The response's http status code.

Response Lenses

deleteColumnStatisticsForPartitionResponse_httpStatus :: Lens' DeleteColumnStatisticsForPartitionResponse Int

The response's http status code.