Network.AWS.Glue.DeleteRegistry

Description

Delete the entire registry including schema and all of its versions. To get the status of the delete operation, you can call the GetRegistry API after the asynchronous call. Deleting a registry will deactivate all online operations for the registry such as the UpdateRegistry, CreateSchema, UpdateSchema, and RegisterSchemaVersion APIs.

Creating a Request

data DeleteRegistry

See: newDeleteRegistry smart constructor.

Constructors

DeleteRegistry'

Fields

registryId :: RegistryId

This is a wrapper structure that may contain the registry name and Amazon Resource Name (ARN).

newDeleteRegistry

Arguments

:: RegistryId
$sel:registryId:DeleteRegistry'

-> DeleteRegistry

Create a value of DeleteRegistry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryId:DeleteRegistry', deleteRegistry_registryId - This is a wrapper structure that may contain the registry name and Amazon Resource Name (ARN).

Request Lenses

deleteRegistry_registryId :: Lens' DeleteRegistry RegistryId

This is a wrapper structure that may contain the registry name and Amazon Resource Name (ARN).

Destructuring the Response

data DeleteRegistryResponse

See: newDeleteRegistryResponse smart constructor.

Constructors

DeleteRegistryResponse'

Fields

status :: Maybe RegistryStatus

The status of the registry. A successful operation will return the Deleting status.

registryName :: Maybe Text

The name of the registry being deleted.

registryArn :: Maybe Text

The Amazon Resource Name (ARN) of the registry being deleted.

httpStatus :: Int

The response's http status code.

newDeleteRegistryResponse

Arguments

:: Int
$sel:httpStatus:DeleteRegistryResponse'

-> DeleteRegistryResponse

Create a value of DeleteRegistryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DeleteRegistryResponse', deleteRegistryResponse_status - The status of the registry. A successful operation will return the Deleting status.

$sel:registryName:DeleteRegistryResponse', deleteRegistryResponse_registryName - The name of the registry being deleted.

$sel:registryArn:DeleteRegistryResponse', deleteRegistryResponse_registryArn - The Amazon Resource Name (ARN) of the registry being deleted.

$sel:httpStatus:DeleteRegistryResponse', deleteRegistryResponse_httpStatus - The response's http status code.

Response Lenses

deleteRegistryResponse_status :: Lens' DeleteRegistryResponse (Maybe RegistryStatus)

The status of the registry. A successful operation will return the Deleting status.

deleteRegistryResponse_registryName :: Lens' DeleteRegistryResponse (Maybe Text)

The name of the registry being deleted.

deleteRegistryResponse_registryArn :: Lens' DeleteRegistryResponse (Maybe Text)

The Amazon Resource Name (ARN) of the registry being deleted.

deleteRegistryResponse_httpStatus :: Lens' DeleteRegistryResponse Int

The response's http status code.