Network.AWS.Glue.DeleteSchema

Description

Deletes the entire schema set, including the schema set and all of its versions. To get the status of the delete operation, you can call GetSchema API after the asynchronous call. Deleting a registry will deactivate all online operations for the schema, such as the GetSchemaByDefinition, and RegisterSchemaVersion APIs.

Creating a Request

data DeleteSchema

See: newDeleteSchema smart constructor.

Constructors

DeleteSchema'

Fields

schemaId :: SchemaId

This is a wrapper structure that may contain the schema name and Amazon Resource Name (ARN).

newDeleteSchema

Arguments

:: SchemaId
$sel:schemaId:DeleteSchema'

-> DeleteSchema

Create a value of DeleteSchema with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaId:DeleteSchema', deleteSchema_schemaId - This is a wrapper structure that may contain the schema name and Amazon Resource Name (ARN).

Request Lenses

deleteSchema_schemaId :: Lens' DeleteSchema SchemaId

This is a wrapper structure that may contain the schema name and Amazon Resource Name (ARN).

Destructuring the Response

data DeleteSchemaResponse

See: newDeleteSchemaResponse smart constructor.

Constructors

DeleteSchemaResponse'

Fields

status :: Maybe SchemaStatus

The status of the schema.

schemaArn :: Maybe Text

The Amazon Resource Name (ARN) of the schema being deleted.

schemaName :: Maybe Text

The name of the schema being deleted.

httpStatus :: Int

The response's http status code.

newDeleteSchemaResponse

Arguments

:: Int
$sel:httpStatus:DeleteSchemaResponse'

-> DeleteSchemaResponse

Create a value of DeleteSchemaResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DeleteSchemaResponse', deleteSchemaResponse_status - The status of the schema.

$sel:schemaArn:DeleteSchemaResponse', deleteSchemaResponse_schemaArn - The Amazon Resource Name (ARN) of the schema being deleted.

$sel:schemaName:DeleteSchemaResponse', deleteSchemaResponse_schemaName - The name of the schema being deleted.

$sel:httpStatus:DeleteSchemaResponse', deleteSchemaResponse_httpStatus - The response's http status code.

Response Lenses

deleteSchemaResponse_status :: Lens' DeleteSchemaResponse (Maybe SchemaStatus)

The status of the schema.

deleteSchemaResponse_schemaArn :: Lens' DeleteSchemaResponse (Maybe Text)

The Amazon Resource Name (ARN) of the schema being deleted.

deleteSchemaResponse_schemaName :: Lens' DeleteSchemaResponse (Maybe Text)

The name of the schema being deleted.

deleteSchemaResponse_httpStatus :: Lens' DeleteSchemaResponse Int

The response's http status code.