Network.AWS.Glue.GetBlueprint

Description

Retrieves the details of a blueprint.

Creating a Request

data GetBlueprint

See: newGetBlueprint smart constructor.

Constructors

GetBlueprint'

Fields

includeBlueprint :: Maybe Bool

Specifies whether or not to include the blueprint in the response.

includeParameterSpec :: Maybe Bool

Specifies whether or not to include the parameter specification.

name :: Text

The name of the blueprint.

newGetBlueprint

Arguments

:: Text
$sel:name:GetBlueprint'

-> GetBlueprint

Create a value of GetBlueprint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeBlueprint:GetBlueprint', getBlueprint_includeBlueprint - Specifies whether or not to include the blueprint in the response.

$sel:includeParameterSpec:GetBlueprint', getBlueprint_includeParameterSpec - Specifies whether or not to include the parameter specification.

$sel:name:GetBlueprint', getBlueprint_name - The name of the blueprint.

Request Lenses

getBlueprint_includeBlueprint :: Lens' GetBlueprint (Maybe Bool)

Specifies whether or not to include the blueprint in the response.

getBlueprint_includeParameterSpec :: Lens' GetBlueprint (Maybe Bool)

Specifies whether or not to include the parameter specification.

getBlueprint_name :: Lens' GetBlueprint Text

The name of the blueprint.

Destructuring the Response

data GetBlueprintResponse

See: newGetBlueprintResponse smart constructor.

Constructors

GetBlueprintResponse'

Fields

blueprint :: Maybe Blueprint

Returns a Blueprint object.

httpStatus :: Int

The response's http status code.

newGetBlueprintResponse

Arguments

:: Int
$sel:httpStatus:GetBlueprintResponse'

-> GetBlueprintResponse

Create a value of GetBlueprintResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blueprint:GetBlueprintResponse', getBlueprintResponse_blueprint - Returns a Blueprint object.

$sel:httpStatus:GetBlueprintResponse', getBlueprintResponse_httpStatus - The response's http status code.

Response Lenses

getBlueprintResponse_blueprint :: Lens' GetBlueprintResponse (Maybe Blueprint)

Returns a Blueprint object.

getBlueprintResponse_httpStatus :: Lens' GetBlueprintResponse Int

The response's http status code.