Network.AWS.Glue.GetColumnStatisticsForPartition

Description

Retrieves partition statistics of columns.

The Identity and Access Management (IAM) permission required for this operation is GetPartition.

Creating a Request

data GetColumnStatisticsForPartition

See: newGetColumnStatisticsForPartition smart constructor.

Constructors

GetColumnStatisticsForPartition'

Fields

catalogId :: Maybe Text

The ID of the Data Catalog where the partitions in question reside. If none is supplied, the Amazon Web Services account ID is used by default.

databaseName :: Text

The name of the catalog database where the partitions reside.

tableName :: Text

The name of the partitions' table.

partitionValues :: [Text]

A list of partition values identifying the partition.

columnNames :: [Text]

A list of the column names.

newGetColumnStatisticsForPartition

Arguments

:: Text
$sel:databaseName:GetColumnStatisticsForPartition'

-> Text
$sel:tableName:GetColumnStatisticsForPartition'

-> GetColumnStatisticsForPartition

Create a value of GetColumnStatisticsForPartition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:GetColumnStatisticsForPartition', getColumnStatisticsForPartition_catalogId - The ID of the Data Catalog where the partitions in question reside. If none is supplied, the Amazon Web Services account ID is used by default.

$sel:databaseName:GetColumnStatisticsForPartition', getColumnStatisticsForPartition_databaseName - The name of the catalog database where the partitions reside.

$sel:tableName:GetColumnStatisticsForPartition', getColumnStatisticsForPartition_tableName - The name of the partitions' table.

$sel:partitionValues:GetColumnStatisticsForPartition', getColumnStatisticsForPartition_partitionValues - A list of partition values identifying the partition.

$sel:columnNames:GetColumnStatisticsForPartition', getColumnStatisticsForPartition_columnNames - A list of the column names.

Request Lenses

getColumnStatisticsForPartition_catalogId :: Lens' GetColumnStatisticsForPartition (Maybe Text)

The ID of the Data Catalog where the partitions in question reside. If none is supplied, the Amazon Web Services account ID is used by default.

getColumnStatisticsForPartition_databaseName :: Lens' GetColumnStatisticsForPartition Text

The name of the catalog database where the partitions reside.

getColumnStatisticsForPartition_tableName :: Lens' GetColumnStatisticsForPartition Text

The name of the partitions' table.

getColumnStatisticsForPartition_partitionValues :: Lens' GetColumnStatisticsForPartition [Text]

A list of partition values identifying the partition.

getColumnStatisticsForPartition_columnNames :: Lens' GetColumnStatisticsForPartition [Text]

A list of the column names.

Destructuring the Response

data GetColumnStatisticsForPartitionResponse

See: newGetColumnStatisticsForPartitionResponse smart constructor.

Constructors

GetColumnStatisticsForPartitionResponse'

Fields

columnStatisticsList :: Maybe [ColumnStatistics]

List of ColumnStatistics that failed to be retrieved.

errors :: Maybe [ColumnError]

Error occurred during retrieving column statistics data.

httpStatus :: Int

The response's http status code.

newGetColumnStatisticsForPartitionResponse

Arguments

:: Int
$sel:httpStatus:GetColumnStatisticsForPartitionResponse'

-> GetColumnStatisticsForPartitionResponse

Create a value of GetColumnStatisticsForPartitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:columnStatisticsList:GetColumnStatisticsForPartitionResponse', getColumnStatisticsForPartitionResponse_columnStatisticsList - List of ColumnStatistics that failed to be retrieved.

$sel:errors:GetColumnStatisticsForPartitionResponse', getColumnStatisticsForPartitionResponse_errors - Error occurred during retrieving column statistics data.

$sel:httpStatus:GetColumnStatisticsForPartitionResponse', getColumnStatisticsForPartitionResponse_httpStatus - The response's http status code.

Response Lenses

getColumnStatisticsForPartitionResponse_columnStatisticsList :: Lens' GetColumnStatisticsForPartitionResponse (Maybe [ColumnStatistics])

List of ColumnStatistics that failed to be retrieved.

getColumnStatisticsForPartitionResponse_errors :: Lens' GetColumnStatisticsForPartitionResponse (Maybe [ColumnError])

Error occurred during retrieving column statistics data.

getColumnStatisticsForPartitionResponse_httpStatus :: Lens' GetColumnStatisticsForPartitionResponse Int

The response's http status code.