Network.AWS.Glue.GetConnections

Description

Retrieves a list of connection definitions from the Data Catalog.

This operation returns paginated results.

Creating a Request

data GetConnections

See: newGetConnections smart constructor.

Constructors

GetConnections'

Fields

nextToken :: Maybe Text

A continuation token, if this is a continuation call.

maxResults :: Maybe Natural

The maximum number of connections to return in one response.

catalogId :: Maybe Text

The ID of the Data Catalog in which the connections reside. If none is provided, the Amazon Web Services account ID is used by default.

hidePassword :: Maybe Bool

Allows you to retrieve the connection metadata without returning the password. For instance, the AWS Glue console uses this flag to retrieve the connection, and does not display the password. Set this parameter when the caller might not have permission to use the KMS key to decrypt the password, but it does have permission to access the rest of the connection properties.

filter' :: Maybe GetConnectionsFilter

A filter that controls which connections are returned.

newGetConnections :: GetConnections

Create a value of GetConnections with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetConnections', getConnections_nextToken - A continuation token, if this is a continuation call.

$sel:maxResults:GetConnections', getConnections_maxResults - The maximum number of connections to return in one response.

$sel:catalogId:GetConnections', getConnections_catalogId - The ID of the Data Catalog in which the connections reside. If none is provided, the Amazon Web Services account ID is used by default.

$sel:hidePassword:GetConnections', getConnections_hidePassword - Allows you to retrieve the connection metadata without returning the password. For instance, the AWS Glue console uses this flag to retrieve the connection, and does not display the password. Set this parameter when the caller might not have permission to use the KMS key to decrypt the password, but it does have permission to access the rest of the connection properties.

$sel:filter':GetConnections', getConnections_filter - A filter that controls which connections are returned.

Request Lenses

getConnections_nextToken :: Lens' GetConnections (Maybe Text)

A continuation token, if this is a continuation call.

getConnections_maxResults :: Lens' GetConnections (Maybe Natural)

The maximum number of connections to return in one response.

getConnections_catalogId :: Lens' GetConnections (Maybe Text)

The ID of the Data Catalog in which the connections reside. If none is provided, the Amazon Web Services account ID is used by default.

getConnections_hidePassword :: Lens' GetConnections (Maybe Bool)

Allows you to retrieve the connection metadata without returning the password. For instance, the AWS Glue console uses this flag to retrieve the connection, and does not display the password. Set this parameter when the caller might not have permission to use the KMS key to decrypt the password, but it does have permission to access the rest of the connection properties.

getConnections_filter :: Lens' GetConnections (Maybe GetConnectionsFilter)

A filter that controls which connections are returned.

Destructuring the Response

data GetConnectionsResponse

See: newGetConnectionsResponse smart constructor.

Constructors

GetConnectionsResponse'

Fields

nextToken :: Maybe Text

A continuation token, if the list of connections returned does not include the last of the filtered connections.

connectionList :: Maybe [Connection]

A list of requested connection definitions.

httpStatus :: Int

The response's http status code.

newGetConnectionsResponse

Arguments

:: Int
$sel:httpStatus:GetConnectionsResponse'

-> GetConnectionsResponse

Create a value of GetConnectionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetConnections', getConnectionsResponse_nextToken - A continuation token, if the list of connections returned does not include the last of the filtered connections.

$sel:connectionList:GetConnectionsResponse', getConnectionsResponse_connectionList - A list of requested connection definitions.

$sel:httpStatus:GetConnectionsResponse', getConnectionsResponse_httpStatus - The response's http status code.

Response Lenses

getConnectionsResponse_nextToken :: Lens' GetConnectionsResponse (Maybe Text)

A continuation token, if the list of connections returned does not include the last of the filtered connections.

getConnectionsResponse_connectionList :: Lens' GetConnectionsResponse (Maybe [Connection])

A list of requested connection definitions.

getConnectionsResponse_httpStatus :: Lens' GetConnectionsResponse Int

The response's http status code.