Network.AWS.Glue.GetJobs

Description

Retrieves all current job definitions.

This operation returns paginated results.

Creating a Request

data GetJobs

See: newGetJobs smart constructor.

Constructors

GetJobs'

Fields

nextToken :: Maybe Text

A continuation token, if this is a continuation call.

maxResults :: Maybe Natural

The maximum size of the response.

newGetJobs :: GetJobs

Create a value of GetJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetJobs', getJobs_nextToken - A continuation token, if this is a continuation call.

$sel:maxResults:GetJobs', getJobs_maxResults - The maximum size of the response.

Request Lenses

getJobs_nextToken :: Lens' GetJobs (Maybe Text)

A continuation token, if this is a continuation call.

getJobs_maxResults :: Lens' GetJobs (Maybe Natural)

The maximum size of the response.

Destructuring the Response

data GetJobsResponse

See: newGetJobsResponse smart constructor.

Constructors

GetJobsResponse'

Fields

nextToken :: Maybe Text

A continuation token, if not all job definitions have yet been returned.

jobs :: Maybe [Job]

A list of job definitions.

httpStatus :: Int

The response's http status code.

newGetJobsResponse

Arguments

:: Int
$sel:httpStatus:GetJobsResponse'

-> GetJobsResponse

Create a value of GetJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetJobs', getJobsResponse_nextToken - A continuation token, if not all job definitions have yet been returned.

$sel:jobs:GetJobsResponse', getJobsResponse_jobs - A list of job definitions.

$sel:httpStatus:GetJobsResponse', getJobsResponse_httpStatus - The response's http status code.

Response Lenses

getJobsResponse_nextToken :: Lens' GetJobsResponse (Maybe Text)

A continuation token, if not all job definitions have yet been returned.

getJobsResponse_jobs :: Lens' GetJobsResponse (Maybe [Job])

A list of job definitions.

getJobsResponse_httpStatus :: Lens' GetJobsResponse Int

The response's http status code.