Network.AWS.Glue.GetMLTaskRun

Description

Gets details for a specific task run on a machine learning transform. Machine learning task runs are asynchronous tasks that Glue runs on your behalf as part of various machine learning workflows. You can check the stats of any task run by calling GetMLTaskRun with the TaskRunID and its parent transform's TransformID.

Creating a Request

data GetMLTaskRun

See: newGetMLTaskRun smart constructor.

Constructors

GetMLTaskRun'

Fields

transformId :: Text

The unique identifier of the machine learning transform.

taskRunId :: Text

The unique identifier of the task run.

newGetMLTaskRun

Arguments

:: Text
$sel:transformId:GetMLTaskRun'

-> Text
$sel:taskRunId:GetMLTaskRun'

-> GetMLTaskRun

Create a value of GetMLTaskRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transformId:GetMLTaskRun', getMLTaskRun_transformId - The unique identifier of the machine learning transform.

$sel:taskRunId:GetMLTaskRun', getMLTaskRun_taskRunId - The unique identifier of the task run.

Request Lenses

getMLTaskRun_transformId :: Lens' GetMLTaskRun Text

The unique identifier of the machine learning transform.

getMLTaskRun_taskRunId :: Lens' GetMLTaskRun Text

The unique identifier of the task run.

Destructuring the Response

data GetMLTaskRunResponse

See: newGetMLTaskRunResponse smart constructor.

Constructors

GetMLTaskRunResponse'

Fields

executionTime :: Maybe Int

The amount of time (in seconds) that the task run consumed resources.

status :: Maybe TaskStatusType

The status for this task run.

taskRunId :: Maybe Text

The unique run identifier associated with this run.

transformId :: Maybe Text

The unique identifier of the task run.

errorString :: Maybe Text

The error strings that are associated with the task run.

lastModifiedOn :: Maybe POSIX

The date and time when this task run was last modified.

completedOn :: Maybe POSIX

The date and time when this task run was completed.

logGroupName :: Maybe Text

The names of the log groups that are associated with the task run.

properties :: Maybe TaskRunProperties

The list of properties that are associated with the task run.

startedOn :: Maybe POSIX

The date and time when this task run started.

httpStatus :: Int

The response's http status code.

newGetMLTaskRunResponse

Arguments

:: Int
$sel:httpStatus:GetMLTaskRunResponse'

-> GetMLTaskRunResponse

Create a value of GetMLTaskRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:executionTime:GetMLTaskRunResponse', getMLTaskRunResponse_executionTime - The amount of time (in seconds) that the task run consumed resources.

$sel:status:GetMLTaskRunResponse', getMLTaskRunResponse_status - The status for this task run.

$sel:taskRunId:GetMLTaskRun', getMLTaskRunResponse_taskRunId - The unique run identifier associated with this run.

$sel:transformId:GetMLTaskRun', getMLTaskRunResponse_transformId - The unique identifier of the task run.

$sel:errorString:GetMLTaskRunResponse', getMLTaskRunResponse_errorString - The error strings that are associated with the task run.

$sel:lastModifiedOn:GetMLTaskRunResponse', getMLTaskRunResponse_lastModifiedOn - The date and time when this task run was last modified.

$sel:completedOn:GetMLTaskRunResponse', getMLTaskRunResponse_completedOn - The date and time when this task run was completed.

$sel:logGroupName:GetMLTaskRunResponse', getMLTaskRunResponse_logGroupName - The names of the log groups that are associated with the task run.

$sel:properties:GetMLTaskRunResponse', getMLTaskRunResponse_properties - The list of properties that are associated with the task run.

$sel:startedOn:GetMLTaskRunResponse', getMLTaskRunResponse_startedOn - The date and time when this task run started.

$sel:httpStatus:GetMLTaskRunResponse', getMLTaskRunResponse_httpStatus - The response's http status code.

Response Lenses

getMLTaskRunResponse_executionTime :: Lens' GetMLTaskRunResponse (Maybe Int)

The amount of time (in seconds) that the task run consumed resources.

getMLTaskRunResponse_status :: Lens' GetMLTaskRunResponse (Maybe TaskStatusType)

The status for this task run.

getMLTaskRunResponse_taskRunId :: Lens' GetMLTaskRunResponse (Maybe Text)

The unique run identifier associated with this run.

getMLTaskRunResponse_transformId :: Lens' GetMLTaskRunResponse (Maybe Text)

The unique identifier of the task run.

getMLTaskRunResponse_errorString :: Lens' GetMLTaskRunResponse (Maybe Text)

The error strings that are associated with the task run.

getMLTaskRunResponse_lastModifiedOn :: Lens' GetMLTaskRunResponse (Maybe UTCTime)

The date and time when this task run was last modified.

getMLTaskRunResponse_completedOn :: Lens' GetMLTaskRunResponse (Maybe UTCTime)

The date and time when this task run was completed.

getMLTaskRunResponse_logGroupName :: Lens' GetMLTaskRunResponse (Maybe Text)

The names of the log groups that are associated with the task run.

getMLTaskRunResponse_properties :: Lens' GetMLTaskRunResponse (Maybe TaskRunProperties)

The list of properties that are associated with the task run.

getMLTaskRunResponse_startedOn :: Lens' GetMLTaskRunResponse (Maybe UTCTime)

The date and time when this task run started.

getMLTaskRunResponse_httpStatus :: Lens' GetMLTaskRunResponse Int

The response's http status code.