Network.AWS.Glue.GetMLTaskRuns

Description

Gets a list of runs for a machine learning transform. Machine learning task runs are asynchronous tasks that Glue runs on your behalf as part of various machine learning workflows. You can get a sortable, filterable list of machine learning task runs by calling GetMLTaskRuns with their parent transform's TransformID and other optional parameters as documented in this section.

This operation returns a list of historic runs and must be paginated.

Creating a Request

data GetMLTaskRuns

See: newGetMLTaskRuns smart constructor.

Constructors

GetMLTaskRuns'

Fields

nextToken :: Maybe Text

A token for pagination of the results. The default is empty.

sort :: Maybe TaskRunSortCriteria

The sorting criteria, in the TaskRunSortCriteria structure, for the task run.

filter' :: Maybe TaskRunFilterCriteria

The filter criteria, in the TaskRunFilterCriteria structure, for the task run.

maxResults :: Maybe Natural

The maximum number of results to return.

transformId :: Text

The unique identifier of the machine learning transform.

newGetMLTaskRuns

Arguments

:: Text
$sel:transformId:GetMLTaskRuns'

-> GetMLTaskRuns

Create a value of GetMLTaskRuns with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetMLTaskRuns', getMLTaskRuns_nextToken - A token for pagination of the results. The default is empty.

$sel:sort:GetMLTaskRuns', getMLTaskRuns_sort - The sorting criteria, in the TaskRunSortCriteria structure, for the task run.

$sel:filter':GetMLTaskRuns', getMLTaskRuns_filter - The filter criteria, in the TaskRunFilterCriteria structure, for the task run.

$sel:maxResults:GetMLTaskRuns', getMLTaskRuns_maxResults - The maximum number of results to return.

$sel:transformId:GetMLTaskRuns', getMLTaskRuns_transformId - The unique identifier of the machine learning transform.

Request Lenses

getMLTaskRuns_nextToken :: Lens' GetMLTaskRuns (Maybe Text)

A token for pagination of the results. The default is empty.

getMLTaskRuns_sort :: Lens' GetMLTaskRuns (Maybe TaskRunSortCriteria)

The sorting criteria, in the TaskRunSortCriteria structure, for the task run.

getMLTaskRuns_filter :: Lens' GetMLTaskRuns (Maybe TaskRunFilterCriteria)

The filter criteria, in the TaskRunFilterCriteria structure, for the task run.

getMLTaskRuns_maxResults :: Lens' GetMLTaskRuns (Maybe Natural)

The maximum number of results to return.

getMLTaskRuns_transformId :: Lens' GetMLTaskRuns Text

The unique identifier of the machine learning transform.

Destructuring the Response

data GetMLTaskRunsResponse

See: newGetMLTaskRunsResponse smart constructor.

Constructors

GetMLTaskRunsResponse'

Fields

nextToken :: Maybe Text

A pagination token, if more results are available.

taskRuns :: Maybe [TaskRun]

A list of task runs that are associated with the transform.

httpStatus :: Int

The response's http status code.

newGetMLTaskRunsResponse

Arguments

:: Int
$sel:httpStatus:GetMLTaskRunsResponse'

-> GetMLTaskRunsResponse

Create a value of GetMLTaskRunsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetMLTaskRuns', getMLTaskRunsResponse_nextToken - A pagination token, if more results are available.

$sel:taskRuns:GetMLTaskRunsResponse', getMLTaskRunsResponse_taskRuns - A list of task runs that are associated with the transform.

$sel:httpStatus:GetMLTaskRunsResponse', getMLTaskRunsResponse_httpStatus - The response's http status code.

Response Lenses

getMLTaskRunsResponse_nextToken :: Lens' GetMLTaskRunsResponse (Maybe Text)

A pagination token, if more results are available.

getMLTaskRunsResponse_taskRuns :: Lens' GetMLTaskRunsResponse (Maybe [TaskRun])

A list of task runs that are associated with the transform.

getMLTaskRunsResponse_httpStatus :: Lens' GetMLTaskRunsResponse Int

The response's http status code.