Network.AWS.Glue.GetMapping

Description

Creates mappings.

Creating a Request

data GetMapping

See: newGetMapping smart constructor.

Constructors

GetMapping'

Fields

sinks :: Maybe [CatalogEntry]

A list of target tables.

location :: Maybe Location

Parameters for the mapping.

source :: CatalogEntry

Specifies the source table.

newGetMapping

Arguments

:: CatalogEntry
$sel:source:GetMapping'

-> GetMapping

Create a value of GetMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sinks:GetMapping', getMapping_sinks - A list of target tables.

$sel:location:GetMapping', getMapping_location - Parameters for the mapping.

$sel:source:GetMapping', getMapping_source - Specifies the source table.

Request Lenses

getMapping_sinks :: Lens' GetMapping (Maybe [CatalogEntry])

A list of target tables.

getMapping_location :: Lens' GetMapping (Maybe Location)

Parameters for the mapping.

getMapping_source :: Lens' GetMapping CatalogEntry

Specifies the source table.

Destructuring the Response

data GetMappingResponse

See: newGetMappingResponse smart constructor.

Constructors

GetMappingResponse'

Fields

httpStatus :: Int

The response's http status code.

mapping :: [MappingEntry]

A list of mappings to the specified targets.

newGetMappingResponse

Arguments

:: Int
$sel:httpStatus:GetMappingResponse'

-> GetMappingResponse

Create a value of GetMappingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetMappingResponse', getMappingResponse_httpStatus - The response's http status code.

$sel:mapping:GetMappingResponse', getMappingResponse_mapping - A list of mappings to the specified targets.

Response Lenses

getMappingResponse_httpStatus :: Lens' GetMappingResponse Int

The response's http status code.

getMappingResponse_mapping :: Lens' GetMappingResponse [MappingEntry]

A list of mappings to the specified targets.