Network.AWS.Glue.GetPartitions

Description

Retrieves information about the partitions in a table.

This operation returns paginated results.

Creating a Request

data GetPartitions

See: newGetPartitions smart constructor.

Constructors

GetPartitions'

Fields

nextToken :: Maybe Text

A continuation token, if this is not the first call to retrieve these partitions.

maxResults :: Maybe Natural

The maximum number of partitions to return in a single response.

catalogId :: Maybe Text

The ID of the Data Catalog where the partitions in question reside. If none is provided, the Amazon Web Services account ID is used by default.

segment :: Maybe Segment

The segment of the table's partitions to scan in this request.

excludeColumnSchema :: Maybe Bool

expression :: Maybe Text

An expression that filters the partitions to be returned.

The expression uses SQL syntax similar to the SQL WHERE filter clause. The SQL statement parser JSQLParser parses the expression.

Operators: The following are the operators that you can use in the Expression API call:

=

Checks whether the values of the two operands are equal; if yes, then the condition becomes true.
Example: Assume 'variable a' holds 10 and 'variable b' holds 20.

(a = b) is not true.

< >

Checks whether the values of two operands are equal; if the values are not equal, then the condition becomes true.
Example: (a < > b) is true.

>

Checks whether the value of the left operand is greater than the value of the right operand; if yes, then the condition becomes true.
Example: (a > b) is not true.

<

Checks whether the value of the left operand is less than the value of the right operand; if yes, then the condition becomes true.
Example: (a < b) is true.

>=

Checks whether the value of the left operand is greater than or equal to the value of the right operand; if yes, then the condition becomes true.
Example: (a >= b) is not true.

<=

Checks whether the value of the left operand is less than or equal to the value of the right operand; if yes, then the condition becomes true.
Example: (a <= b) is true.

AND, OR, IN, BETWEEN, LIKE, NOT, IS NULL

Logical operators.

Supported Partition Key Types: The following are the supported partition keys.

string

date

timestamp

int

bigint

long

tinyint

smallint

decimal

If an type is encountered that is not valid, an exception is thrown.

The following list shows the valid operators on each type. When you define a crawler, the partitionKey type is created as a STRING, to be compatible with the catalog partitions.

Sample API Call:

databaseName :: Text

The name of the catalog database where the partitions reside.

tableName :: Text

The name of the partitions' table.

newGetPartitions

Arguments

:: Text
$sel:databaseName:GetPartitions'

-> Text
$sel:tableName:GetPartitions'

-> GetPartitions

Create a value of GetPartitions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetPartitions', getPartitions_nextToken - A continuation token, if this is not the first call to retrieve these partitions.

$sel:maxResults:GetPartitions', getPartitions_maxResults - The maximum number of partitions to return in a single response.

$sel:catalogId:GetPartitions', getPartitions_catalogId - The ID of the Data Catalog where the partitions in question reside. If none is provided, the Amazon Web Services account ID is used by default.

$sel:segment:GetPartitions', getPartitions_segment - The segment of the table's partitions to scan in this request.

$sel:excludeColumnSchema:GetPartitions', getPartitions_excludeColumnSchema - Undocumented member.

$sel:expression:GetPartitions', getPartitions_expression - An expression that filters the partitions to be returned.

The expression uses SQL syntax similar to the SQL WHERE filter clause. The SQL statement parser JSQLParser parses the expression.

Operators: The following are the operators that you can use in the Expression API call:

=

Checks whether the values of the two operands are equal; if yes, then the condition becomes true.
Example: Assume 'variable a' holds 10 and 'variable b' holds 20.

(a = b) is not true.

< >

Checks whether the values of two operands are equal; if the values are not equal, then the condition becomes true.
Example: (a < > b) is true.

>

Checks whether the value of the left operand is greater than the value of the right operand; if yes, then the condition becomes true.
Example: (a > b) is not true.

<

Checks whether the value of the left operand is less than the value of the right operand; if yes, then the condition becomes true.
Example: (a < b) is true.

>=

Checks whether the value of the left operand is greater than or equal to the value of the right operand; if yes, then the condition becomes true.
Example: (a >= b) is not true.

<=

Checks whether the value of the left operand is less than or equal to the value of the right operand; if yes, then the condition becomes true.
Example: (a <= b) is true.

AND, OR, IN, BETWEEN, LIKE, NOT, IS NULL

Logical operators.

Supported Partition Key Types: The following are the supported partition keys.

string

date

timestamp

int

bigint

long

tinyint

smallint

decimal

If an type is encountered that is not valid, an exception is thrown.

The following list shows the valid operators on each type. When you define a crawler, the partitionKey type is created as a STRING, to be compatible with the catalog partitions.

Sample API Call:

$sel:databaseName:GetPartitions', getPartitions_databaseName - The name of the catalog database where the partitions reside.

$sel:tableName:GetPartitions', getPartitions_tableName - The name of the partitions' table.

Request Lenses

getPartitions_nextToken :: Lens' GetPartitions (Maybe Text)

A continuation token, if this is not the first call to retrieve these partitions.

getPartitions_maxResults :: Lens' GetPartitions (Maybe Natural)

The maximum number of partitions to return in a single response.

getPartitions_catalogId :: Lens' GetPartitions (Maybe Text)

The ID of the Data Catalog where the partitions in question reside. If none is provided, the Amazon Web Services account ID is used by default.

getPartitions_segment :: Lens' GetPartitions (Maybe Segment)

The segment of the table's partitions to scan in this request.

getPartitions_excludeColumnSchema :: Lens' GetPartitions (Maybe Bool)

Undocumented member.

getPartitions_expression :: Lens' GetPartitions (Maybe Text)

An expression that filters the partitions to be returned.

The expression uses SQL syntax similar to the SQL WHERE filter clause. The SQL statement parser JSQLParser parses the expression.

Operators: The following are the operators that you can use in the Expression API call:

=

Checks whether the values of the two operands are equal; if yes, then the condition becomes true.
Example: Assume 'variable a' holds 10 and 'variable b' holds 20.

(a = b) is not true.

< >

Checks whether the values of two operands are equal; if the values are not equal, then the condition becomes true.
Example: (a < > b) is true.

>

Checks whether the value of the left operand is greater than the value of the right operand; if yes, then the condition becomes true.
Example: (a > b) is not true.

<

Checks whether the value of the left operand is less than the value of the right operand; if yes, then the condition becomes true.
Example: (a < b) is true.

>=

Checks whether the value of the left operand is greater than or equal to the value of the right operand; if yes, then the condition becomes true.
Example: (a >= b) is not true.

<=

Checks whether the value of the left operand is less than or equal to the value of the right operand; if yes, then the condition becomes true.
Example: (a <= b) is true.

AND, OR, IN, BETWEEN, LIKE, NOT, IS NULL

Logical operators.

Supported Partition Key Types: The following are the supported partition keys.

string

date

timestamp

int

bigint

long

tinyint

smallint

decimal

If an type is encountered that is not valid, an exception is thrown.

The following list shows the valid operators on each type. When you define a crawler, the partitionKey type is created as a STRING, to be compatible with the catalog partitions.

Sample API Call:

getPartitions_databaseName :: Lens' GetPartitions Text

The name of the catalog database where the partitions reside.

getPartitions_tableName :: Lens' GetPartitions Text

The name of the partitions' table.

Destructuring the Response

data GetPartitionsResponse

See: newGetPartitionsResponse smart constructor.

Constructors

GetPartitionsResponse'

Fields

nextToken :: Maybe Text

A continuation token, if the returned list of partitions does not include the last one.

partitions :: Maybe [Partition]

A list of requested partitions.

httpStatus :: Int

The response's http status code.

newGetPartitionsResponse

Arguments

:: Int
$sel:httpStatus:GetPartitionsResponse'

-> GetPartitionsResponse

Create a value of GetPartitionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetPartitions', getPartitionsResponse_nextToken - A continuation token, if the returned list of partitions does not include the last one.

$sel:partitions:GetPartitionsResponse', getPartitionsResponse_partitions - A list of requested partitions.

$sel:httpStatus:GetPartitionsResponse', getPartitionsResponse_httpStatus - The response's http status code.

Response Lenses

getPartitionsResponse_nextToken :: Lens' GetPartitionsResponse (Maybe Text)

A continuation token, if the returned list of partitions does not include the last one.

getPartitionsResponse_partitions :: Lens' GetPartitionsResponse (Maybe [Partition])

A list of requested partitions.

getPartitionsResponse_httpStatus :: Lens' GetPartitionsResponse Int

The response's http status code.