Network.AWS.Glue.GetResourcePolicy

Description

Retrieves a specified resource policy.

Creating a Request

data GetResourcePolicy

See: newGetResourcePolicy smart constructor.

Constructors

GetResourcePolicy'

Fields

resourceArn :: Maybe Text

The ARN of the Glue resource for which to retrieve the resource policy. If not supplied, the Data Catalog resource policy is returned. Use GetResourcePolicies to view all existing resource policies. For more information see Specifying Glue Resource ARNs.

newGetResourcePolicy :: GetResourcePolicy

Create a value of GetResourcePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:GetResourcePolicy', getResourcePolicy_resourceArn - The ARN of the Glue resource for which to retrieve the resource policy. If not supplied, the Data Catalog resource policy is returned. Use GetResourcePolicies to view all existing resource policies. For more information see Specifying Glue Resource ARNs.

Request Lenses

getResourcePolicy_resourceArn :: Lens' GetResourcePolicy (Maybe Text)

The ARN of the Glue resource for which to retrieve the resource policy. If not supplied, the Data Catalog resource policy is returned. Use GetResourcePolicies to view all existing resource policies. For more information see Specifying Glue Resource ARNs.

Destructuring the Response

data GetResourcePolicyResponse

See: newGetResourcePolicyResponse smart constructor.

Constructors

GetResourcePolicyResponse'

Fields

policyInJson :: Maybe Text

Contains the requested policy document, in JSON format.

updateTime :: Maybe POSIX

The date and time at which the policy was last updated.

createTime :: Maybe POSIX

The date and time at which the policy was created.

policyHash :: Maybe Text

Contains the hash value associated with this policy.

httpStatus :: Int

The response's http status code.

newGetResourcePolicyResponse

Arguments

:: Int
$sel:httpStatus:GetResourcePolicyResponse'

-> GetResourcePolicyResponse

Create a value of GetResourcePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyInJson:GetResourcePolicyResponse', getResourcePolicyResponse_policyInJson - Contains the requested policy document, in JSON format.

$sel:updateTime:GetResourcePolicyResponse', getResourcePolicyResponse_updateTime - The date and time at which the policy was last updated.

$sel:createTime:GetResourcePolicyResponse', getResourcePolicyResponse_createTime - The date and time at which the policy was created.

$sel:policyHash:GetResourcePolicyResponse', getResourcePolicyResponse_policyHash - Contains the hash value associated with this policy.

$sel:httpStatus:GetResourcePolicyResponse', getResourcePolicyResponse_httpStatus - The response's http status code.

Response Lenses

getResourcePolicyResponse_policyInJson :: Lens' GetResourcePolicyResponse (Maybe Text)

Contains the requested policy document, in JSON format.

getResourcePolicyResponse_updateTime :: Lens' GetResourcePolicyResponse (Maybe UTCTime)

The date and time at which the policy was last updated.

getResourcePolicyResponse_createTime :: Lens' GetResourcePolicyResponse (Maybe UTCTime)

The date and time at which the policy was created.

getResourcePolicyResponse_policyHash :: Lens' GetResourcePolicyResponse (Maybe Text)

Contains the hash value associated with this policy.

getResourcePolicyResponse_httpStatus :: Lens' GetResourcePolicyResponse Int

The response's http status code.