Network.AWS.Glue.GetSchema

Description

Describes the specified schema in detail.

Creating a Request

data GetSchema

See: newGetSchema smart constructor.

Constructors

GetSchema'

Fields

schemaId :: SchemaId

This is a wrapper structure to contain schema identity fields. The structure contains:

SchemaId$SchemaArn: The Amazon Resource Name (ARN) of the schema. Either SchemaArn or SchemaName and RegistryName has to be provided.

SchemaId$SchemaName: The name of the schema. Either SchemaArn or SchemaName and RegistryName has to be provided.

newGetSchema

Arguments

:: SchemaId
$sel:schemaId:GetSchema'

-> GetSchema

Create a value of GetSchema with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaId:GetSchema', getSchema_schemaId - This is a wrapper structure to contain schema identity fields. The structure contains:

SchemaId$SchemaArn: The Amazon Resource Name (ARN) of the schema. Either SchemaArn or SchemaName and RegistryName has to be provided.

SchemaId$SchemaName: The name of the schema. Either SchemaArn or SchemaName and RegistryName has to be provided.

Request Lenses

getSchema_schemaId :: Lens' GetSchema SchemaId

This is a wrapper structure to contain schema identity fields. The structure contains:

SchemaId$SchemaArn: The Amazon Resource Name (ARN) of the schema. Either SchemaArn or SchemaName and RegistryName has to be provided.

SchemaId$SchemaName: The name of the schema. Either SchemaArn or SchemaName and RegistryName has to be provided.

Destructuring the Response

data GetSchemaResponse

See: newGetSchemaResponse smart constructor.

Constructors

GetSchemaResponse'

Fields

registryName :: Maybe Text

The name of the registry.

createdTime :: Maybe Text

The date and time the schema was created.

schemaStatus :: Maybe SchemaStatus

The status of the schema.

registryArn :: Maybe Text

The Amazon Resource Name (ARN) of the registry.

latestSchemaVersion :: Maybe Natural

The latest version of the schema associated with the returned schema definition.

dataFormat :: Maybe DataFormat

The data format of the schema definition. Currently AVRO and JSON are supported.

schemaCheckpoint :: Maybe Natural

The version number of the checkpoint (the last time the compatibility mode was changed).

schemaName :: Maybe Text

The name of the schema.

schemaArn :: Maybe Text

The Amazon Resource Name (ARN) of the schema.

nextSchemaVersion :: Maybe Natural

The next version of the schema associated with the returned schema definition.

updatedTime :: Maybe Text

The date and time the schema was updated.

description :: Maybe Text

A description of schema if specified when created

compatibility :: Maybe Compatibility

The compatibility mode of the schema.

httpStatus :: Int

The response's http status code.

newGetSchemaResponse

Arguments

:: Int
$sel:httpStatus:GetSchemaResponse'

-> GetSchemaResponse

Create a value of GetSchemaResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryName:GetSchemaResponse', getSchemaResponse_registryName - The name of the registry.

$sel:createdTime:GetSchemaResponse', getSchemaResponse_createdTime - The date and time the schema was created.

$sel:schemaStatus:GetSchemaResponse', getSchemaResponse_schemaStatus - The status of the schema.

$sel:registryArn:GetSchemaResponse', getSchemaResponse_registryArn - The Amazon Resource Name (ARN) of the registry.

$sel:latestSchemaVersion:GetSchemaResponse', getSchemaResponse_latestSchemaVersion - The latest version of the schema associated with the returned schema definition.

$sel:dataFormat:GetSchemaResponse', getSchemaResponse_dataFormat - The data format of the schema definition. Currently AVRO and JSON are supported.

$sel:schemaCheckpoint:GetSchemaResponse', getSchemaResponse_schemaCheckpoint - The version number of the checkpoint (the last time the compatibility mode was changed).

$sel:schemaName:GetSchemaResponse', getSchemaResponse_schemaName - The name of the schema.

$sel:schemaArn:GetSchemaResponse', getSchemaResponse_schemaArn - The Amazon Resource Name (ARN) of the schema.

$sel:nextSchemaVersion:GetSchemaResponse', getSchemaResponse_nextSchemaVersion - The next version of the schema associated with the returned schema definition.

$sel:updatedTime:GetSchemaResponse', getSchemaResponse_updatedTime - The date and time the schema was updated.

$sel:description:GetSchemaResponse', getSchemaResponse_description - A description of schema if specified when created

$sel:compatibility:GetSchemaResponse', getSchemaResponse_compatibility - The compatibility mode of the schema.

$sel:httpStatus:GetSchemaResponse', getSchemaResponse_httpStatus - The response's http status code.

Response Lenses

getSchemaResponse_registryName :: Lens' GetSchemaResponse (Maybe Text)

The name of the registry.

getSchemaResponse_createdTime :: Lens' GetSchemaResponse (Maybe Text)

The date and time the schema was created.

getSchemaResponse_schemaStatus :: Lens' GetSchemaResponse (Maybe SchemaStatus)

The status of the schema.

getSchemaResponse_registryArn :: Lens' GetSchemaResponse (Maybe Text)

The Amazon Resource Name (ARN) of the registry.

getSchemaResponse_latestSchemaVersion :: Lens' GetSchemaResponse (Maybe Natural)

The latest version of the schema associated with the returned schema definition.

getSchemaResponse_dataFormat :: Lens' GetSchemaResponse (Maybe DataFormat)

The data format of the schema definition. Currently AVRO and JSON are supported.

getSchemaResponse_schemaCheckpoint :: Lens' GetSchemaResponse (Maybe Natural)

The version number of the checkpoint (the last time the compatibility mode was changed).

getSchemaResponse_schemaName :: Lens' GetSchemaResponse (Maybe Text)

The name of the schema.

getSchemaResponse_schemaArn :: Lens' GetSchemaResponse (Maybe Text)

The Amazon Resource Name (ARN) of the schema.

getSchemaResponse_nextSchemaVersion :: Lens' GetSchemaResponse (Maybe Natural)

The next version of the schema associated with the returned schema definition.

getSchemaResponse_updatedTime :: Lens' GetSchemaResponse (Maybe Text)

The date and time the schema was updated.

getSchemaResponse_description :: Lens' GetSchemaResponse (Maybe Text)

A description of schema if specified when created

getSchemaResponse_compatibility :: Lens' GetSchemaResponse (Maybe Compatibility)

The compatibility mode of the schema.

getSchemaResponse_httpStatus :: Lens' GetSchemaResponse Int

The response's http status code.