Network.AWS.Glue.GetSchemaVersion

Description

Get the specified schema by its unique ID assigned when a version of the schema is created or registered. Schema versions in Deleted status will not be included in the results.

Creating a Request

data GetSchemaVersion

See: newGetSchemaVersion smart constructor.

Constructors

GetSchemaVersion'

Fields

schemaVersionId :: Maybe Text

The SchemaVersionId of the schema version. This field is required for fetching by schema ID. Either this or the SchemaId wrapper has to be provided.

schemaVersionNumber :: Maybe SchemaVersionNumber

The version number of the schema.

schemaId :: Maybe SchemaId

This is a wrapper structure to contain schema identity fields. The structure contains:

SchemaId$SchemaArn: The Amazon Resource Name (ARN) of the schema. Either SchemaArn or SchemaName and RegistryName has to be provided.

SchemaId$SchemaName: The name of the schema. Either SchemaArn or SchemaName and RegistryName has to be provided.

newGetSchemaVersion :: GetSchemaVersion

Create a value of GetSchemaVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaVersionId:GetSchemaVersion', getSchemaVersion_schemaVersionId - The SchemaVersionId of the schema version. This field is required for fetching by schema ID. Either this or the SchemaId wrapper has to be provided.

$sel:schemaVersionNumber:GetSchemaVersion', getSchemaVersion_schemaVersionNumber - The version number of the schema.

$sel:schemaId:GetSchemaVersion', getSchemaVersion_schemaId - This is a wrapper structure to contain schema identity fields. The structure contains:

SchemaId$SchemaArn: The Amazon Resource Name (ARN) of the schema. Either SchemaArn or SchemaName and RegistryName has to be provided.

SchemaId$SchemaName: The name of the schema. Either SchemaArn or SchemaName and RegistryName has to be provided.

Request Lenses

getSchemaVersion_schemaVersionId :: Lens' GetSchemaVersion (Maybe Text)

The SchemaVersionId of the schema version. This field is required for fetching by schema ID. Either this or the SchemaId wrapper has to be provided.

getSchemaVersion_schemaVersionNumber :: Lens' GetSchemaVersion (Maybe SchemaVersionNumber)

The version number of the schema.

getSchemaVersion_schemaId :: Lens' GetSchemaVersion (Maybe SchemaId)

This is a wrapper structure to contain schema identity fields. The structure contains:

SchemaId$SchemaArn: The Amazon Resource Name (ARN) of the schema. Either SchemaArn or SchemaName and RegistryName has to be provided.

SchemaId$SchemaName: The name of the schema. Either SchemaArn or SchemaName and RegistryName has to be provided.

Destructuring the Response

data GetSchemaVersionResponse

See: newGetSchemaVersionResponse smart constructor.

Constructors

GetSchemaVersionResponse'

Fields

status :: Maybe SchemaVersionStatus

The status of the schema version.

schemaDefinition :: Maybe Text

The schema definition for the schema ID.

schemaArn :: Maybe Text

The Amazon Resource Name (ARN) of the schema.

schemaVersionId :: Maybe Text

The SchemaVersionId of the schema version.

dataFormat :: Maybe DataFormat

The data format of the schema definition. Currently AVRO and JSON are supported.

createdTime :: Maybe Text

The date and time the schema version was created.

versionNumber :: Maybe Natural

The version number of the schema.

httpStatus :: Int

The response's http status code.

newGetSchemaVersionResponse

Arguments

:: Int
$sel:httpStatus:GetSchemaVersionResponse'

-> GetSchemaVersionResponse

Create a value of GetSchemaVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetSchemaVersionResponse', getSchemaVersionResponse_status - The status of the schema version.

$sel:schemaDefinition:GetSchemaVersionResponse', getSchemaVersionResponse_schemaDefinition - The schema definition for the schema ID.

$sel:schemaArn:GetSchemaVersionResponse', getSchemaVersionResponse_schemaArn - The Amazon Resource Name (ARN) of the schema.

$sel:schemaVersionId:GetSchemaVersion', getSchemaVersionResponse_schemaVersionId - The SchemaVersionId of the schema version.

$sel:dataFormat:GetSchemaVersionResponse', getSchemaVersionResponse_dataFormat - The data format of the schema definition. Currently AVRO and JSON are supported.

$sel:createdTime:GetSchemaVersionResponse', getSchemaVersionResponse_createdTime - The date and time the schema version was created.

$sel:versionNumber:GetSchemaVersionResponse', getSchemaVersionResponse_versionNumber - The version number of the schema.

$sel:httpStatus:GetSchemaVersionResponse', getSchemaVersionResponse_httpStatus - The response's http status code.

Response Lenses

getSchemaVersionResponse_status :: Lens' GetSchemaVersionResponse (Maybe SchemaVersionStatus)

The status of the schema version.

getSchemaVersionResponse_schemaDefinition :: Lens' GetSchemaVersionResponse (Maybe Text)

The schema definition for the schema ID.

getSchemaVersionResponse_schemaArn :: Lens' GetSchemaVersionResponse (Maybe Text)

The Amazon Resource Name (ARN) of the schema.

getSchemaVersionResponse_schemaVersionId :: Lens' GetSchemaVersionResponse (Maybe Text)

The SchemaVersionId of the schema version.

getSchemaVersionResponse_dataFormat :: Lens' GetSchemaVersionResponse (Maybe DataFormat)

The data format of the schema definition. Currently AVRO and JSON are supported.

getSchemaVersionResponse_createdTime :: Lens' GetSchemaVersionResponse (Maybe Text)

The date and time the schema version was created.

getSchemaVersionResponse_versionNumber :: Lens' GetSchemaVersionResponse (Maybe Natural)

The version number of the schema.

getSchemaVersionResponse_httpStatus :: Lens' GetSchemaVersionResponse Int

The response's http status code.