Network.AWS.Glue.GetTables

Description

Retrieves the definitions of some or all of the tables in a given Database.

This operation returns paginated results.

Creating a Request

data GetTables

See: newGetTables smart constructor.

Constructors

GetTables'

Fields

nextToken :: Maybe Text

A continuation token, included if this is a continuation call.

maxResults :: Maybe Natural

The maximum number of tables to return in a single response.

catalogId :: Maybe Text

The ID of the Data Catalog where the tables reside. If none is provided, the Amazon Web Services account ID is used by default.

expression :: Maybe Text

A regular expression pattern. If present, only those tables whose names match the pattern are returned.

databaseName :: Text

The database in the catalog whose tables to list. For Hive compatibility, this name is entirely lowercase.

newGetTables

Arguments

:: Text
$sel:databaseName:GetTables'

-> GetTables

Create a value of GetTables with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetTables', getTables_nextToken - A continuation token, included if this is a continuation call.

$sel:maxResults:GetTables', getTables_maxResults - The maximum number of tables to return in a single response.

$sel:catalogId:GetTables', getTables_catalogId - The ID of the Data Catalog where the tables reside. If none is provided, the Amazon Web Services account ID is used by default.

$sel:expression:GetTables', getTables_expression - A regular expression pattern. If present, only those tables whose names match the pattern are returned.

$sel:databaseName:GetTables', getTables_databaseName - The database in the catalog whose tables to list. For Hive compatibility, this name is entirely lowercase.

Request Lenses

getTables_nextToken :: Lens' GetTables (Maybe Text)

A continuation token, included if this is a continuation call.

getTables_maxResults :: Lens' GetTables (Maybe Natural)

The maximum number of tables to return in a single response.

getTables_catalogId :: Lens' GetTables (Maybe Text)

The ID of the Data Catalog where the tables reside. If none is provided, the Amazon Web Services account ID is used by default.

getTables_expression :: Lens' GetTables (Maybe Text)

A regular expression pattern. If present, only those tables whose names match the pattern are returned.

getTables_databaseName :: Lens' GetTables Text

The database in the catalog whose tables to list. For Hive compatibility, this name is entirely lowercase.

Destructuring the Response

data GetTablesResponse

See: newGetTablesResponse smart constructor.

Constructors

GetTablesResponse'

Fields

nextToken :: Maybe Text

A continuation token, present if the current list segment is not the last.

tableList :: Maybe [Table]

A list of the requested Table objects.

httpStatus :: Int

The response's http status code.

newGetTablesResponse

Arguments

:: Int
$sel:httpStatus:GetTablesResponse'

-> GetTablesResponse

Create a value of GetTablesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetTables', getTablesResponse_nextToken - A continuation token, present if the current list segment is not the last.

$sel:tableList:GetTablesResponse', getTablesResponse_tableList - A list of the requested Table objects.

$sel:httpStatus:GetTablesResponse', getTablesResponse_httpStatus - The response's http status code.

Response Lenses

getTablesResponse_nextToken :: Lens' GetTablesResponse (Maybe Text)

A continuation token, present if the current list segment is not the last.

getTablesResponse_tableList :: Lens' GetTablesResponse (Maybe [Table])

A list of the requested Table objects.

getTablesResponse_httpStatus :: Lens' GetTablesResponse Int

The response's http status code.