Network.AWS.Glue.GetUserDefinedFunction

Description

Retrieves a specified function definition from the Data Catalog.

Creating a Request

data GetUserDefinedFunction

See: newGetUserDefinedFunction smart constructor.

Constructors

GetUserDefinedFunction'

Fields

catalogId :: Maybe Text

The ID of the Data Catalog where the function to be retrieved is located. If none is provided, the Amazon Web Services account ID is used by default.

databaseName :: Text

The name of the catalog database where the function is located.

functionName :: Text

The name of the function.

newGetUserDefinedFunction

Arguments

:: Text
$sel:databaseName:GetUserDefinedFunction'

-> Text
$sel:functionName:GetUserDefinedFunction'

-> GetUserDefinedFunction

Create a value of GetUserDefinedFunction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:GetUserDefinedFunction', getUserDefinedFunction_catalogId - The ID of the Data Catalog where the function to be retrieved is located. If none is provided, the Amazon Web Services account ID is used by default.

$sel:databaseName:GetUserDefinedFunction', getUserDefinedFunction_databaseName - The name of the catalog database where the function is located.

$sel:functionName:GetUserDefinedFunction', getUserDefinedFunction_functionName - The name of the function.

Request Lenses

getUserDefinedFunction_catalogId :: Lens' GetUserDefinedFunction (Maybe Text)

The ID of the Data Catalog where the function to be retrieved is located. If none is provided, the Amazon Web Services account ID is used by default.

getUserDefinedFunction_databaseName :: Lens' GetUserDefinedFunction Text

The name of the catalog database where the function is located.

getUserDefinedFunction_functionName :: Lens' GetUserDefinedFunction Text

The name of the function.

Destructuring the Response

data GetUserDefinedFunctionResponse

See: newGetUserDefinedFunctionResponse smart constructor.

Constructors

GetUserDefinedFunctionResponse'

Fields

userDefinedFunction :: Maybe UserDefinedFunction

The requested function definition.

httpStatus :: Int

The response's http status code.

newGetUserDefinedFunctionResponse

Arguments

:: Int
$sel:httpStatus:GetUserDefinedFunctionResponse'

-> GetUserDefinedFunctionResponse

Create a value of GetUserDefinedFunctionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userDefinedFunction:GetUserDefinedFunctionResponse', getUserDefinedFunctionResponse_userDefinedFunction - The requested function definition.

$sel:httpStatus:GetUserDefinedFunctionResponse', getUserDefinedFunctionResponse_httpStatus - The response's http status code.

Response Lenses

getUserDefinedFunctionResponse_userDefinedFunction :: Lens' GetUserDefinedFunctionResponse (Maybe UserDefinedFunction)

The requested function definition.

getUserDefinedFunctionResponse_httpStatus :: Lens' GetUserDefinedFunctionResponse Int

The response's http status code.