Network.AWS.Glue.GetWorkflow

Description

Retrieves resource metadata for a workflow.

Creating a Request

data GetWorkflow

See: newGetWorkflow smart constructor.

Constructors

GetWorkflow'

Fields

includeGraph :: Maybe Bool

Specifies whether to include a graph when returning the workflow resource metadata.

name :: Text

The name of the workflow to retrieve.

newGetWorkflow

Arguments

:: Text
$sel:name:GetWorkflow'

-> GetWorkflow

Create a value of GetWorkflow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeGraph:GetWorkflow', getWorkflow_includeGraph - Specifies whether to include a graph when returning the workflow resource metadata.

$sel:name:GetWorkflow', getWorkflow_name - The name of the workflow to retrieve.

Request Lenses

getWorkflow_includeGraph :: Lens' GetWorkflow (Maybe Bool)

Specifies whether to include a graph when returning the workflow resource metadata.

getWorkflow_name :: Lens' GetWorkflow Text

The name of the workflow to retrieve.

Destructuring the Response

data GetWorkflowResponse

See: newGetWorkflowResponse smart constructor.

Constructors

GetWorkflowResponse'

Fields

workflow :: Maybe Workflow

The resource metadata for the workflow.

httpStatus :: Int

The response's http status code.

newGetWorkflowResponse

Arguments

:: Int
$sel:httpStatus:GetWorkflowResponse'

-> GetWorkflowResponse

Create a value of GetWorkflowResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workflow:GetWorkflowResponse', getWorkflowResponse_workflow - The resource metadata for the workflow.

$sel:httpStatus:GetWorkflowResponse', getWorkflowResponse_httpStatus - The response's http status code.

Response Lenses

getWorkflowResponse_workflow :: Lens' GetWorkflowResponse (Maybe Workflow)

The resource metadata for the workflow.

getWorkflowResponse_httpStatus :: Lens' GetWorkflowResponse Int

The response's http status code.