Network.AWS.Glue.GetWorkflowRuns

Description

Retrieves metadata for all runs of a given workflow.

Creating a Request

data GetWorkflowRuns

See: newGetWorkflowRuns smart constructor.

Constructors

GetWorkflowRuns'

Fields

includeGraph :: Maybe Bool

Specifies whether to include the workflow graph in response or not.

nextToken :: Maybe Text

The maximum size of the response.

maxResults :: Maybe Natural

The maximum number of workflow runs to be included in the response.

name :: Text

Name of the workflow whose metadata of runs should be returned.

newGetWorkflowRuns

Arguments

:: Text
$sel:name:GetWorkflowRuns'

-> GetWorkflowRuns

Create a value of GetWorkflowRuns with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeGraph:GetWorkflowRuns', getWorkflowRuns_includeGraph - Specifies whether to include the workflow graph in response or not.

$sel:nextToken:GetWorkflowRuns', getWorkflowRuns_nextToken - The maximum size of the response.

$sel:maxResults:GetWorkflowRuns', getWorkflowRuns_maxResults - The maximum number of workflow runs to be included in the response.

$sel:name:GetWorkflowRuns', getWorkflowRuns_name - Name of the workflow whose metadata of runs should be returned.

Request Lenses

getWorkflowRuns_includeGraph :: Lens' GetWorkflowRuns (Maybe Bool)

Specifies whether to include the workflow graph in response or not.

getWorkflowRuns_nextToken :: Lens' GetWorkflowRuns (Maybe Text)

The maximum size of the response.

getWorkflowRuns_maxResults :: Lens' GetWorkflowRuns (Maybe Natural)

The maximum number of workflow runs to be included in the response.

getWorkflowRuns_name :: Lens' GetWorkflowRuns Text

Name of the workflow whose metadata of runs should be returned.

Destructuring the Response

data GetWorkflowRunsResponse

See: newGetWorkflowRunsResponse smart constructor.

Constructors

GetWorkflowRunsResponse'

Fields

runs :: Maybe (NonEmpty WorkflowRun)

A list of workflow run metadata objects.

nextToken :: Maybe Text

A continuation token, if not all requested workflow runs have been returned.

httpStatus :: Int

The response's http status code.

newGetWorkflowRunsResponse

Arguments

:: Int
$sel:httpStatus:GetWorkflowRunsResponse'

-> GetWorkflowRunsResponse

Create a value of GetWorkflowRunsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:runs:GetWorkflowRunsResponse', getWorkflowRunsResponse_runs - A list of workflow run metadata objects.

$sel:nextToken:GetWorkflowRuns', getWorkflowRunsResponse_nextToken - A continuation token, if not all requested workflow runs have been returned.

$sel:httpStatus:GetWorkflowRunsResponse', getWorkflowRunsResponse_httpStatus - The response's http status code.

Response Lenses

getWorkflowRunsResponse_runs :: Lens' GetWorkflowRunsResponse (Maybe (NonEmpty WorkflowRun))

A list of workflow run metadata objects.

getWorkflowRunsResponse_nextToken :: Lens' GetWorkflowRunsResponse (Maybe Text)

A continuation token, if not all requested workflow runs have been returned.

getWorkflowRunsResponse_httpStatus :: Lens' GetWorkflowRunsResponse Int

The response's http status code.