Network.AWS.Glue.ListSchemaVersions

Description

Returns a list of schema versions that you have created, with minimal information. Schema versions in Deleted status will not be included in the results. Empty results will be returned if there are no schema versions available.

This operation returns paginated results.

Creating a Request

data ListSchemaVersions

See: newListSchemaVersions smart constructor.

Constructors

ListSchemaVersions'

Fields

nextToken :: Maybe Text

A continuation token, if this is a continuation call.

maxResults :: Maybe Natural

Maximum number of results required per page. If the value is not supplied, this will be defaulted to 25 per page.

schemaId :: SchemaId

This is a wrapper structure to contain schema identity fields. The structure contains:

SchemaId$SchemaArn: The Amazon Resource Name (ARN) of the schema. Either SchemaArn or SchemaName and RegistryName has to be provided.

SchemaId$SchemaName: The name of the schema. Either SchemaArn or SchemaName and RegistryName has to be provided.

newListSchemaVersions

Arguments

:: SchemaId
$sel:schemaId:ListSchemaVersions'

-> ListSchemaVersions

Create a value of ListSchemaVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSchemaVersions', listSchemaVersions_nextToken - A continuation token, if this is a continuation call.

$sel:maxResults:ListSchemaVersions', listSchemaVersions_maxResults - Maximum number of results required per page. If the value is not supplied, this will be defaulted to 25 per page.

$sel:schemaId:ListSchemaVersions', listSchemaVersions_schemaId - This is a wrapper structure to contain schema identity fields. The structure contains:

SchemaId$SchemaArn: The Amazon Resource Name (ARN) of the schema. Either SchemaArn or SchemaName and RegistryName has to be provided.

SchemaId$SchemaName: The name of the schema. Either SchemaArn or SchemaName and RegistryName has to be provided.

Request Lenses

listSchemaVersions_nextToken :: Lens' ListSchemaVersions (Maybe Text)

A continuation token, if this is a continuation call.

listSchemaVersions_maxResults :: Lens' ListSchemaVersions (Maybe Natural)

Maximum number of results required per page. If the value is not supplied, this will be defaulted to 25 per page.

listSchemaVersions_schemaId :: Lens' ListSchemaVersions SchemaId

This is a wrapper structure to contain schema identity fields. The structure contains:

SchemaId$SchemaArn: The Amazon Resource Name (ARN) of the schema. Either SchemaArn or SchemaName and RegistryName has to be provided.

SchemaId$SchemaName: The name of the schema. Either SchemaArn or SchemaName and RegistryName has to be provided.

Destructuring the Response

data ListSchemaVersionsResponse

See: newListSchemaVersionsResponse smart constructor.

Constructors

ListSchemaVersionsResponse'

Fields

schemas :: Maybe [SchemaVersionListItem]

An array of SchemaVersionList objects containing details of each schema version.

nextToken :: Maybe Text

A continuation token for paginating the returned list of tokens, returned if the current segment of the list is not the last.

httpStatus :: Int

The response's http status code.

newListSchemaVersionsResponse

Arguments

:: Int
$sel:httpStatus:ListSchemaVersionsResponse'

-> ListSchemaVersionsResponse

Create a value of ListSchemaVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemas:ListSchemaVersionsResponse', listSchemaVersionsResponse_schemas - An array of SchemaVersionList objects containing details of each schema version.

$sel:nextToken:ListSchemaVersions', listSchemaVersionsResponse_nextToken - A continuation token for paginating the returned list of tokens, returned if the current segment of the list is not the last.

$sel:httpStatus:ListSchemaVersionsResponse', listSchemaVersionsResponse_httpStatus - The response's http status code.

Response Lenses

listSchemaVersionsResponse_schemas :: Lens' ListSchemaVersionsResponse (Maybe [SchemaVersionListItem])

An array of SchemaVersionList objects containing details of each schema version.

listSchemaVersionsResponse_nextToken :: Lens' ListSchemaVersionsResponse (Maybe Text)

A continuation token for paginating the returned list of tokens, returned if the current segment of the list is not the last.

listSchemaVersionsResponse_httpStatus :: Lens' ListSchemaVersionsResponse Int

The response's http status code.