Network.AWS.Glue.ListTriggers

Description

Retrieves the names of all trigger resources in this Amazon Web Services account, or the resources with the specified tag. This operation allows you to see which resources are available in your account, and their names.

This operation takes the optional Tags field, which you can use as a filter on the response so that tagged resources can be retrieved as a group. If you choose to use tags filtering, only resources with the tag are retrieved.

Creating a Request

data ListTriggers

See: newListTriggers smart constructor.

Constructors

ListTriggers'

Fields

nextToken :: Maybe Text

A continuation token, if this is a continuation request.

maxResults :: Maybe Natural

The maximum size of a list to return.

tags :: Maybe (HashMap Text Text)

Specifies to return only these tagged resources.

dependentJobName :: Maybe Text

The name of the job for which to retrieve triggers. The trigger that can start this job is returned. If there is no such trigger, all triggers are returned.

newListTriggers :: ListTriggers

Create a value of ListTriggers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTriggers', listTriggers_nextToken - A continuation token, if this is a continuation request.

$sel:maxResults:ListTriggers', listTriggers_maxResults - The maximum size of a list to return.

$sel:tags:ListTriggers', listTriggers_tags - Specifies to return only these tagged resources.

$sel:dependentJobName:ListTriggers', listTriggers_dependentJobName - The name of the job for which to retrieve triggers. The trigger that can start this job is returned. If there is no such trigger, all triggers are returned.

Request Lenses

listTriggers_nextToken :: Lens' ListTriggers (Maybe Text)

A continuation token, if this is a continuation request.

listTriggers_maxResults :: Lens' ListTriggers (Maybe Natural)

The maximum size of a list to return.

listTriggers_tags :: Lens' ListTriggers (Maybe (HashMap Text Text))

Specifies to return only these tagged resources.

listTriggers_dependentJobName :: Lens' ListTriggers (Maybe Text)

The name of the job for which to retrieve triggers. The trigger that can start this job is returned. If there is no such trigger, all triggers are returned.

Destructuring the Response

data ListTriggersResponse

See: newListTriggersResponse smart constructor.

Constructors

ListTriggersResponse'

Fields

triggerNames :: Maybe [Text]

The names of all triggers in the account, or the triggers with the specified tags.

nextToken :: Maybe Text

A continuation token, if the returned list does not contain the last metric available.

httpStatus :: Int

The response's http status code.

newListTriggersResponse

Arguments

:: Int
$sel:httpStatus:ListTriggersResponse'

-> ListTriggersResponse

Create a value of ListTriggersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:triggerNames:ListTriggersResponse', listTriggersResponse_triggerNames - The names of all triggers in the account, or the triggers with the specified tags.

$sel:nextToken:ListTriggers', listTriggersResponse_nextToken - A continuation token, if the returned list does not contain the last metric available.

$sel:httpStatus:ListTriggersResponse', listTriggersResponse_httpStatus - The response's http status code.

Response Lenses

listTriggersResponse_triggerNames :: Lens' ListTriggersResponse (Maybe [Text])

The names of all triggers in the account, or the triggers with the specified tags.

listTriggersResponse_nextToken :: Lens' ListTriggersResponse (Maybe Text)

A continuation token, if the returned list does not contain the last metric available.

listTriggersResponse_httpStatus :: Lens' ListTriggersResponse Int

The response's http status code.