Network.AWS.Glue.PutSchemaVersionMetadata

Description

Puts the metadata key value pair for a specified schema version ID. A maximum of 10 key value pairs will be allowed per schema version. They can be added over one or more calls.

Creating a Request

data PutSchemaVersionMetadata

See: newPutSchemaVersionMetadata smart constructor.

Constructors

PutSchemaVersionMetadata'

Fields

schemaVersionId :: Maybe Text

The unique version ID of the schema version.

schemaVersionNumber :: Maybe SchemaVersionNumber

The version number of the schema.

schemaId :: Maybe SchemaId

The unique ID for the schema.

metadataKeyValue :: MetadataKeyValuePair

The metadata key's corresponding value.

newPutSchemaVersionMetadata

Arguments

:: MetadataKeyValuePair
$sel:metadataKeyValue:PutSchemaVersionMetadata'

-> PutSchemaVersionMetadata

Create a value of PutSchemaVersionMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaVersionId:PutSchemaVersionMetadata', putSchemaVersionMetadata_schemaVersionId - The unique version ID of the schema version.

$sel:schemaVersionNumber:PutSchemaVersionMetadata', putSchemaVersionMetadata_schemaVersionNumber - The version number of the schema.

$sel:schemaId:PutSchemaVersionMetadata', putSchemaVersionMetadata_schemaId - The unique ID for the schema.

$sel:metadataKeyValue:PutSchemaVersionMetadata', putSchemaVersionMetadata_metadataKeyValue - The metadata key's corresponding value.

Request Lenses

putSchemaVersionMetadata_schemaVersionId :: Lens' PutSchemaVersionMetadata (Maybe Text)

The unique version ID of the schema version.

putSchemaVersionMetadata_schemaVersionNumber :: Lens' PutSchemaVersionMetadata (Maybe SchemaVersionNumber)

The version number of the schema.

putSchemaVersionMetadata_schemaId :: Lens' PutSchemaVersionMetadata (Maybe SchemaId)

The unique ID for the schema.

putSchemaVersionMetadata_metadataKeyValue :: Lens' PutSchemaVersionMetadata MetadataKeyValuePair

The metadata key's corresponding value.

Destructuring the Response

data PutSchemaVersionMetadataResponse

See: newPutSchemaVersionMetadataResponse smart constructor.

Constructors

PutSchemaVersionMetadataResponse'

Fields

schemaArn :: Maybe Text

The Amazon Resource Name (ARN) for the schema.

latestVersion :: Maybe Bool

The latest version of the schema.

schemaVersionId :: Maybe Text

The unique version ID of the schema version.

registryName :: Maybe Text

The name for the registry.

metadataKey :: Maybe Text

The metadata key.

versionNumber :: Maybe Natural

The version number of the schema.

metadataValue :: Maybe Text

The value of the metadata key.

schemaName :: Maybe Text

The name for the schema.

httpStatus :: Int

The response's http status code.

newPutSchemaVersionMetadataResponse

Arguments

:: Int
$sel:httpStatus:PutSchemaVersionMetadataResponse'

-> PutSchemaVersionMetadataResponse

Create a value of PutSchemaVersionMetadataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaArn:PutSchemaVersionMetadataResponse', putSchemaVersionMetadataResponse_schemaArn - The Amazon Resource Name (ARN) for the schema.

$sel:latestVersion:PutSchemaVersionMetadataResponse', putSchemaVersionMetadataResponse_latestVersion - The latest version of the schema.

$sel:schemaVersionId:PutSchemaVersionMetadata', putSchemaVersionMetadataResponse_schemaVersionId - The unique version ID of the schema version.

$sel:registryName:PutSchemaVersionMetadataResponse', putSchemaVersionMetadataResponse_registryName - The name for the registry.

$sel:metadataKey:PutSchemaVersionMetadataResponse', putSchemaVersionMetadataResponse_metadataKey - The metadata key.

$sel:versionNumber:PutSchemaVersionMetadataResponse', putSchemaVersionMetadataResponse_versionNumber - The version number of the schema.

$sel:metadataValue:PutSchemaVersionMetadataResponse', putSchemaVersionMetadataResponse_metadataValue - The value of the metadata key.

$sel:schemaName:PutSchemaVersionMetadataResponse', putSchemaVersionMetadataResponse_schemaName - The name for the schema.

$sel:httpStatus:PutSchemaVersionMetadataResponse', putSchemaVersionMetadataResponse_httpStatus - The response's http status code.

Response Lenses

putSchemaVersionMetadataResponse_schemaArn :: Lens' PutSchemaVersionMetadataResponse (Maybe Text)

The Amazon Resource Name (ARN) for the schema.

putSchemaVersionMetadataResponse_latestVersion :: Lens' PutSchemaVersionMetadataResponse (Maybe Bool)

The latest version of the schema.

putSchemaVersionMetadataResponse_schemaVersionId :: Lens' PutSchemaVersionMetadataResponse (Maybe Text)

The unique version ID of the schema version.

putSchemaVersionMetadataResponse_registryName :: Lens' PutSchemaVersionMetadataResponse (Maybe Text)

The name for the registry.

putSchemaVersionMetadataResponse_metadataKey :: Lens' PutSchemaVersionMetadataResponse (Maybe Text)

The metadata key.

putSchemaVersionMetadataResponse_versionNumber :: Lens' PutSchemaVersionMetadataResponse (Maybe Natural)

The version number of the schema.

putSchemaVersionMetadataResponse_metadataValue :: Lens' PutSchemaVersionMetadataResponse (Maybe Text)

The value of the metadata key.

putSchemaVersionMetadataResponse_schemaName :: Lens' PutSchemaVersionMetadataResponse (Maybe Text)

The name for the schema.

putSchemaVersionMetadataResponse_httpStatus :: Lens' PutSchemaVersionMetadataResponse Int

The response's http status code.