Network.AWS.Glue.QuerySchemaVersionMetadata

Description

Queries for the schema version metadata information.

Creating a Request

data QuerySchemaVersionMetadata

See: newQuerySchemaVersionMetadata smart constructor.

Constructors

QuerySchemaVersionMetadata'

Fields

nextToken :: Maybe Text

A continuation token, if this is a continuation call.

maxResults :: Maybe Natural

Maximum number of results required per page. If the value is not supplied, this will be defaulted to 25 per page.

schemaVersionId :: Maybe Text

The unique version ID of the schema version.

schemaVersionNumber :: Maybe SchemaVersionNumber

The version number of the schema.

metadataList :: Maybe [MetadataKeyValuePair]

Search key-value pairs for metadata, if they are not provided all the metadata information will be fetched.

schemaId :: Maybe SchemaId

A wrapper structure that may contain the schema name and Amazon Resource Name (ARN).

newQuerySchemaVersionMetadata :: QuerySchemaVersionMetadata

Create a value of QuerySchemaVersionMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:QuerySchemaVersionMetadata', querySchemaVersionMetadata_nextToken - A continuation token, if this is a continuation call.

$sel:maxResults:QuerySchemaVersionMetadata', querySchemaVersionMetadata_maxResults - Maximum number of results required per page. If the value is not supplied, this will be defaulted to 25 per page.

$sel:schemaVersionId:QuerySchemaVersionMetadata', querySchemaVersionMetadata_schemaVersionId - The unique version ID of the schema version.

$sel:schemaVersionNumber:QuerySchemaVersionMetadata', querySchemaVersionMetadata_schemaVersionNumber - The version number of the schema.

$sel:metadataList:QuerySchemaVersionMetadata', querySchemaVersionMetadata_metadataList - Search key-value pairs for metadata, if they are not provided all the metadata information will be fetched.

$sel:schemaId:QuerySchemaVersionMetadata', querySchemaVersionMetadata_schemaId - A wrapper structure that may contain the schema name and Amazon Resource Name (ARN).

Request Lenses

querySchemaVersionMetadata_nextToken :: Lens' QuerySchemaVersionMetadata (Maybe Text)

A continuation token, if this is a continuation call.

querySchemaVersionMetadata_maxResults :: Lens' QuerySchemaVersionMetadata (Maybe Natural)

Maximum number of results required per page. If the value is not supplied, this will be defaulted to 25 per page.

querySchemaVersionMetadata_schemaVersionId :: Lens' QuerySchemaVersionMetadata (Maybe Text)

The unique version ID of the schema version.

querySchemaVersionMetadata_schemaVersionNumber :: Lens' QuerySchemaVersionMetadata (Maybe SchemaVersionNumber)

The version number of the schema.

querySchemaVersionMetadata_metadataList :: Lens' QuerySchemaVersionMetadata (Maybe [MetadataKeyValuePair])

Search key-value pairs for metadata, if they are not provided all the metadata information will be fetched.

querySchemaVersionMetadata_schemaId :: Lens' QuerySchemaVersionMetadata (Maybe SchemaId)

A wrapper structure that may contain the schema name and Amazon Resource Name (ARN).

Destructuring the Response

data QuerySchemaVersionMetadataResponse

See: newQuerySchemaVersionMetadataResponse smart constructor.

Constructors

QuerySchemaVersionMetadataResponse'

Fields

metadataInfoMap :: Maybe (HashMap Text MetadataInfo)

A map of a metadata key and associated values.

nextToken :: Maybe Text

A continuation token for paginating the returned list of tokens, returned if the current segment of the list is not the last.

schemaVersionId :: Maybe Text

The unique version ID of the schema version.

httpStatus :: Int

The response's http status code.

newQuerySchemaVersionMetadataResponse

Arguments

:: Int
$sel:httpStatus:QuerySchemaVersionMetadataResponse'

-> QuerySchemaVersionMetadataResponse

Create a value of QuerySchemaVersionMetadataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metadataInfoMap:QuerySchemaVersionMetadataResponse', querySchemaVersionMetadataResponse_metadataInfoMap - A map of a metadata key and associated values.

$sel:nextToken:QuerySchemaVersionMetadata', querySchemaVersionMetadataResponse_nextToken - A continuation token for paginating the returned list of tokens, returned if the current segment of the list is not the last.

$sel:schemaVersionId:QuerySchemaVersionMetadata', querySchemaVersionMetadataResponse_schemaVersionId - The unique version ID of the schema version.

$sel:httpStatus:QuerySchemaVersionMetadataResponse', querySchemaVersionMetadataResponse_httpStatus - The response's http status code.

Response Lenses

querySchemaVersionMetadataResponse_metadataInfoMap :: Lens' QuerySchemaVersionMetadataResponse (Maybe (HashMap Text MetadataInfo))

A map of a metadata key and associated values.

querySchemaVersionMetadataResponse_nextToken :: Lens' QuerySchemaVersionMetadataResponse (Maybe Text)

A continuation token for paginating the returned list of tokens, returned if the current segment of the list is not the last.

querySchemaVersionMetadataResponse_schemaVersionId :: Lens' QuerySchemaVersionMetadataResponse (Maybe Text)

The unique version ID of the schema version.

querySchemaVersionMetadataResponse_httpStatus :: Lens' QuerySchemaVersionMetadataResponse Int

The response's http status code.