Network.AWS.Glue.RemoveSchemaVersionMetadata

Description

Removes a key value pair from the schema version metadata for the specified schema version ID.

Creating a Request

data RemoveSchemaVersionMetadata

See: newRemoveSchemaVersionMetadata smart constructor.

Constructors

RemoveSchemaVersionMetadata'

Fields

schemaVersionId :: Maybe Text

The unique version ID of the schema version.

schemaVersionNumber :: Maybe SchemaVersionNumber

The version number of the schema.

schemaId :: Maybe SchemaId

A wrapper structure that may contain the schema name and Amazon Resource Name (ARN).

metadataKeyValue :: MetadataKeyValuePair

The value of the metadata key.

newRemoveSchemaVersionMetadata

Arguments

:: MetadataKeyValuePair
$sel:metadataKeyValue:RemoveSchemaVersionMetadata'

-> RemoveSchemaVersionMetadata

Create a value of RemoveSchemaVersionMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaVersionId:RemoveSchemaVersionMetadata', removeSchemaVersionMetadata_schemaVersionId - The unique version ID of the schema version.

$sel:schemaVersionNumber:RemoveSchemaVersionMetadata', removeSchemaVersionMetadata_schemaVersionNumber - The version number of the schema.

$sel:schemaId:RemoveSchemaVersionMetadata', removeSchemaVersionMetadata_schemaId - A wrapper structure that may contain the schema name and Amazon Resource Name (ARN).

$sel:metadataKeyValue:RemoveSchemaVersionMetadata', removeSchemaVersionMetadata_metadataKeyValue - The value of the metadata key.

Request Lenses

removeSchemaVersionMetadata_schemaVersionId :: Lens' RemoveSchemaVersionMetadata (Maybe Text)

The unique version ID of the schema version.

removeSchemaVersionMetadata_schemaVersionNumber :: Lens' RemoveSchemaVersionMetadata (Maybe SchemaVersionNumber)

The version number of the schema.

removeSchemaVersionMetadata_schemaId :: Lens' RemoveSchemaVersionMetadata (Maybe SchemaId)

A wrapper structure that may contain the schema name and Amazon Resource Name (ARN).

removeSchemaVersionMetadata_metadataKeyValue :: Lens' RemoveSchemaVersionMetadata MetadataKeyValuePair

The value of the metadata key.

Destructuring the Response

data RemoveSchemaVersionMetadataResponse

See: newRemoveSchemaVersionMetadataResponse smart constructor.

Constructors

RemoveSchemaVersionMetadataResponse'

Fields

schemaArn :: Maybe Text

The Amazon Resource Name (ARN) of the schema.

latestVersion :: Maybe Bool

The latest version of the schema.

schemaVersionId :: Maybe Text

The version ID for the schema version.

registryName :: Maybe Text

The name of the registry.

metadataKey :: Maybe Text

The metadata key.

versionNumber :: Maybe Natural

The version number of the schema.

metadataValue :: Maybe Text

The value of the metadata key.

schemaName :: Maybe Text

The name of the schema.

httpStatus :: Int

The response's http status code.

newRemoveSchemaVersionMetadataResponse

Arguments

:: Int
$sel:httpStatus:RemoveSchemaVersionMetadataResponse'

-> RemoveSchemaVersionMetadataResponse

Create a value of RemoveSchemaVersionMetadataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaArn:RemoveSchemaVersionMetadataResponse', removeSchemaVersionMetadataResponse_schemaArn - The Amazon Resource Name (ARN) of the schema.

$sel:latestVersion:RemoveSchemaVersionMetadataResponse', removeSchemaVersionMetadataResponse_latestVersion - The latest version of the schema.

$sel:schemaVersionId:RemoveSchemaVersionMetadata', removeSchemaVersionMetadataResponse_schemaVersionId - The version ID for the schema version.

$sel:registryName:RemoveSchemaVersionMetadataResponse', removeSchemaVersionMetadataResponse_registryName - The name of the registry.

$sel:metadataKey:RemoveSchemaVersionMetadataResponse', removeSchemaVersionMetadataResponse_metadataKey - The metadata key.

$sel:versionNumber:RemoveSchemaVersionMetadataResponse', removeSchemaVersionMetadataResponse_versionNumber - The version number of the schema.

$sel:metadataValue:RemoveSchemaVersionMetadataResponse', removeSchemaVersionMetadataResponse_metadataValue - The value of the metadata key.

$sel:schemaName:RemoveSchemaVersionMetadataResponse', removeSchemaVersionMetadataResponse_schemaName - The name of the schema.

$sel:httpStatus:RemoveSchemaVersionMetadataResponse', removeSchemaVersionMetadataResponse_httpStatus - The response's http status code.

Response Lenses

removeSchemaVersionMetadataResponse_schemaArn :: Lens' RemoveSchemaVersionMetadataResponse (Maybe Text)

The Amazon Resource Name (ARN) of the schema.

removeSchemaVersionMetadataResponse_latestVersion :: Lens' RemoveSchemaVersionMetadataResponse (Maybe Bool)

The latest version of the schema.

removeSchemaVersionMetadataResponse_schemaVersionId :: Lens' RemoveSchemaVersionMetadataResponse (Maybe Text)

The version ID for the schema version.

removeSchemaVersionMetadataResponse_registryName :: Lens' RemoveSchemaVersionMetadataResponse (Maybe Text)

The name of the registry.

removeSchemaVersionMetadataResponse_metadataKey :: Lens' RemoveSchemaVersionMetadataResponse (Maybe Text)

The metadata key.

removeSchemaVersionMetadataResponse_versionNumber :: Lens' RemoveSchemaVersionMetadataResponse (Maybe Natural)

The version number of the schema.

removeSchemaVersionMetadataResponse_metadataValue :: Lens' RemoveSchemaVersionMetadataResponse (Maybe Text)

The value of the metadata key.

removeSchemaVersionMetadataResponse_schemaName :: Lens' RemoveSchemaVersionMetadataResponse (Maybe Text)

The name of the schema.

removeSchemaVersionMetadataResponse_httpStatus :: Lens' RemoveSchemaVersionMetadataResponse Int

The response's http status code.