Network.AWS.Glue.StartCrawlerSchedule

Description

Changes the schedule state of the specified crawler to SCHEDULED, unless the crawler is already running or the schedule state is already SCHEDULED.

Creating a Request

data StartCrawlerSchedule

See: newStartCrawlerSchedule smart constructor.

Constructors

StartCrawlerSchedule'

Fields

crawlerName :: Text

Name of the crawler to schedule.

newStartCrawlerSchedule

Arguments

:: Text
$sel:crawlerName:StartCrawlerSchedule'

-> StartCrawlerSchedule

Create a value of StartCrawlerSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:crawlerName:StartCrawlerSchedule', startCrawlerSchedule_crawlerName - Name of the crawler to schedule.

Request Lenses

startCrawlerSchedule_crawlerName :: Lens' StartCrawlerSchedule Text

Name of the crawler to schedule.

Destructuring the Response

data StartCrawlerScheduleResponse

See: newStartCrawlerScheduleResponse smart constructor.

Constructors

StartCrawlerScheduleResponse'

Fields

httpStatus :: Int

The response's http status code.

newStartCrawlerScheduleResponse

Arguments

:: Int
$sel:httpStatus:StartCrawlerScheduleResponse'

-> StartCrawlerScheduleResponse

Create a value of StartCrawlerScheduleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StartCrawlerScheduleResponse', startCrawlerScheduleResponse_httpStatus - The response's http status code.

Response Lenses

startCrawlerScheduleResponse_httpStatus :: Lens' StartCrawlerScheduleResponse Int

The response's http status code.