Network.AWS.Glue.StartImportLabelsTaskRun

Description

Enables you to provide additional labels (examples of truth) to be used to teach the machine learning transform and improve its quality. This API operation is generally used as part of the active learning workflow that starts with the StartMLLabelingSetGenerationTaskRun call and that ultimately results in improving the quality of your machine learning transform.

After the StartMLLabelingSetGenerationTaskRun finishes, Glue machine learning will have generated a series of questions for humans to answer. (Answering these questions is often called 'labeling' in the machine learning workflows). In the case of the FindMatches transform, these questions are of the form, “What is the correct way to group these rows together into groups composed entirely of matching records?” After the labeling process is finished, users upload their answers/labels with a call to StartImportLabelsTaskRun. After StartImportLabelsTaskRun finishes, all future runs of the machine learning transform use the new and improved labels and perform a higher-quality transformation.

By default, StartMLLabelingSetGenerationTaskRun continually learns from and combines all labels that you upload unless you set Replace to true. If you set Replace to true, StartImportLabelsTaskRun deletes and forgets all previously uploaded labels and learns only from the exact set that you upload. Replacing labels can be helpful if you realize that you previously uploaded incorrect labels, and you believe that they are having a negative effect on your transform quality.

You can check on the status of your task run by calling the GetMLTaskRun operation.

Creating a Request

data StartImportLabelsTaskRun

See: newStartImportLabelsTaskRun smart constructor.

Constructors

StartImportLabelsTaskRun'

Fields

replaceAllLabels :: Maybe Bool

Indicates whether to overwrite your existing labels.

transformId :: Text

The unique identifier of the machine learning transform.

inputS3Path :: Text

The Amazon Simple Storage Service (Amazon S3) path from where you import the labels.

newStartImportLabelsTaskRun

Arguments

:: Text
$sel:transformId:StartImportLabelsTaskRun'

-> Text
$sel:inputS3Path:StartImportLabelsTaskRun'

-> StartImportLabelsTaskRun

Create a value of StartImportLabelsTaskRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:replaceAllLabels:StartImportLabelsTaskRun', startImportLabelsTaskRun_replaceAllLabels - Indicates whether to overwrite your existing labels.

$sel:transformId:StartImportLabelsTaskRun', startImportLabelsTaskRun_transformId - The unique identifier of the machine learning transform.

$sel:inputS3Path:StartImportLabelsTaskRun', startImportLabelsTaskRun_inputS3Path - The Amazon Simple Storage Service (Amazon S3) path from where you import the labels.

Request Lenses

startImportLabelsTaskRun_replaceAllLabels :: Lens' StartImportLabelsTaskRun (Maybe Bool)

Indicates whether to overwrite your existing labels.

startImportLabelsTaskRun_transformId :: Lens' StartImportLabelsTaskRun Text

The unique identifier of the machine learning transform.

startImportLabelsTaskRun_inputS3Path :: Lens' StartImportLabelsTaskRun Text

The Amazon Simple Storage Service (Amazon S3) path from where you import the labels.

Destructuring the Response

data StartImportLabelsTaskRunResponse

See: newStartImportLabelsTaskRunResponse smart constructor.

Constructors

StartImportLabelsTaskRunResponse'

Fields

taskRunId :: Maybe Text

The unique identifier for the task run.

httpStatus :: Int

The response's http status code.

newStartImportLabelsTaskRunResponse

Arguments

:: Int
$sel:httpStatus:StartImportLabelsTaskRunResponse'

-> StartImportLabelsTaskRunResponse

Create a value of StartImportLabelsTaskRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskRunId:StartImportLabelsTaskRunResponse', startImportLabelsTaskRunResponse_taskRunId - The unique identifier for the task run.

$sel:httpStatus:StartImportLabelsTaskRunResponse', startImportLabelsTaskRunResponse_httpStatus - The response's http status code.

Response Lenses

startImportLabelsTaskRunResponse_taskRunId :: Lens' StartImportLabelsTaskRunResponse (Maybe Text)

The unique identifier for the task run.

startImportLabelsTaskRunResponse_httpStatus :: Lens' StartImportLabelsTaskRunResponse Int

The response's http status code.