Network.AWS.Glue.StartMLEvaluationTaskRun

Description

Starts a task to estimate the quality of the transform.

When you provide label sets as examples of truth, Glue machine learning uses some of those examples to learn from them. The rest of the labels are used as a test to estimate quality.

Returns a unique identifier for the run. You can call GetMLTaskRun to get more information about the stats of the EvaluationTaskRun.

Creating a Request

data StartMLEvaluationTaskRun

See: newStartMLEvaluationTaskRun smart constructor.

Constructors

StartMLEvaluationTaskRun'

Fields

transformId :: Text

The unique identifier of the machine learning transform.

newStartMLEvaluationTaskRun

Arguments

:: Text
$sel:transformId:StartMLEvaluationTaskRun'

-> StartMLEvaluationTaskRun

Create a value of StartMLEvaluationTaskRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transformId:StartMLEvaluationTaskRun', startMLEvaluationTaskRun_transformId - The unique identifier of the machine learning transform.

Request Lenses

startMLEvaluationTaskRun_transformId :: Lens' StartMLEvaluationTaskRun Text

The unique identifier of the machine learning transform.

Destructuring the Response

data StartMLEvaluationTaskRunResponse

See: newStartMLEvaluationTaskRunResponse smart constructor.

Constructors

StartMLEvaluationTaskRunResponse'

Fields

taskRunId :: Maybe Text

The unique identifier associated with this run.

httpStatus :: Int

The response's http status code.

newStartMLEvaluationTaskRunResponse

Arguments

:: Int
$sel:httpStatus:StartMLEvaluationTaskRunResponse'

-> StartMLEvaluationTaskRunResponse

Create a value of StartMLEvaluationTaskRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskRunId:StartMLEvaluationTaskRunResponse', startMLEvaluationTaskRunResponse_taskRunId - The unique identifier associated with this run.

$sel:httpStatus:StartMLEvaluationTaskRunResponse', startMLEvaluationTaskRunResponse_httpStatus - The response's http status code.

Response Lenses

startMLEvaluationTaskRunResponse_taskRunId :: Lens' StartMLEvaluationTaskRunResponse (Maybe Text)

The unique identifier associated with this run.

startMLEvaluationTaskRunResponse_httpStatus :: Lens' StartMLEvaluationTaskRunResponse Int

The response's http status code.