Network.AWS.Glue.StartMLLabelingSetGenerationTaskRun

Description

Starts the active learning workflow for your machine learning transform to improve the transform's quality by generating label sets and adding labels.

When the StartMLLabelingSetGenerationTaskRun finishes, Glue will have generated a "labeling set" or a set of questions for humans to answer.

In the case of the FindMatches transform, these questions are of the form, “What is the correct way to group these rows together into groups composed entirely of matching records?”

After the labeling process is finished, you can upload your labels with a call to StartImportLabelsTaskRun. After StartImportLabelsTaskRun finishes, all future runs of the machine learning transform will use the new and improved labels and perform a higher-quality transformation.

Creating a Request

data StartMLLabelingSetGenerationTaskRun

See: newStartMLLabelingSetGenerationTaskRun smart constructor.

Constructors

StartMLLabelingSetGenerationTaskRun'

Fields

transformId :: Text

The unique identifier of the machine learning transform.

outputS3Path :: Text

The Amazon Simple Storage Service (Amazon S3) path where you generate the labeling set.

newStartMLLabelingSetGenerationTaskRun

Arguments

:: Text
$sel:transformId:StartMLLabelingSetGenerationTaskRun'

-> Text
$sel:outputS3Path:StartMLLabelingSetGenerationTaskRun'

-> StartMLLabelingSetGenerationTaskRun

Create a value of StartMLLabelingSetGenerationTaskRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transformId:StartMLLabelingSetGenerationTaskRun', startMLLabelingSetGenerationTaskRun_transformId - The unique identifier of the machine learning transform.

$sel:outputS3Path:StartMLLabelingSetGenerationTaskRun', startMLLabelingSetGenerationTaskRun_outputS3Path - The Amazon Simple Storage Service (Amazon S3) path where you generate the labeling set.

Request Lenses

startMLLabelingSetGenerationTaskRun_transformId :: Lens' StartMLLabelingSetGenerationTaskRun Text

The unique identifier of the machine learning transform.

startMLLabelingSetGenerationTaskRun_outputS3Path :: Lens' StartMLLabelingSetGenerationTaskRun Text

The Amazon Simple Storage Service (Amazon S3) path where you generate the labeling set.

Destructuring the Response

data StartMLLabelingSetGenerationTaskRunResponse

See: newStartMLLabelingSetGenerationTaskRunResponse smart constructor.

Constructors

StartMLLabelingSetGenerationTaskRunResponse'

Fields

taskRunId :: Maybe Text

The unique run identifier that is associated with this task run.

httpStatus :: Int

The response's http status code.

newStartMLLabelingSetGenerationTaskRunResponse

Arguments

:: Int
$sel:httpStatus:StartMLLabelingSetGenerationTaskRunResponse'

-> StartMLLabelingSetGenerationTaskRunResponse

Create a value of StartMLLabelingSetGenerationTaskRunResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskRunId:StartMLLabelingSetGenerationTaskRunResponse', startMLLabelingSetGenerationTaskRunResponse_taskRunId - The unique run identifier that is associated with this task run.

$sel:httpStatus:StartMLLabelingSetGenerationTaskRunResponse', startMLLabelingSetGenerationTaskRunResponse_httpStatus - The response's http status code.

Response Lenses

startMLLabelingSetGenerationTaskRunResponse_taskRunId :: Lens' StartMLLabelingSetGenerationTaskRunResponse (Maybe Text)

The unique run identifier that is associated with this task run.

startMLLabelingSetGenerationTaskRunResponse_httpStatus :: Lens' StartMLLabelingSetGenerationTaskRunResponse Int

The response's http status code.