Network.AWS.Glue.StartTrigger

Description

Starts an existing trigger. See Triggering Jobs for information about how different types of trigger are started.

Creating a Request

data StartTrigger

See: newStartTrigger smart constructor.

Constructors

StartTrigger'

Fields

name :: Text

The name of the trigger to start.

newStartTrigger

Arguments

:: Text
$sel:name:StartTrigger'

-> StartTrigger

Create a value of StartTrigger with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:StartTrigger', startTrigger_name - The name of the trigger to start.

Request Lenses

startTrigger_name :: Lens' StartTrigger Text

The name of the trigger to start.

Destructuring the Response

data StartTriggerResponse

See: newStartTriggerResponse smart constructor.

Constructors

StartTriggerResponse'

Fields

name :: Maybe Text

The name of the trigger that was started.

httpStatus :: Int

The response's http status code.

newStartTriggerResponse

Arguments

:: Int
$sel:httpStatus:StartTriggerResponse'

-> StartTriggerResponse

Create a value of StartTriggerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:StartTrigger', startTriggerResponse_name - The name of the trigger that was started.

$sel:httpStatus:StartTriggerResponse', startTriggerResponse_httpStatus - The response's http status code.

Response Lenses

startTriggerResponse_name :: Lens' StartTriggerResponse (Maybe Text)

The name of the trigger that was started.

startTriggerResponse_httpStatus :: Lens' StartTriggerResponse Int

The response's http status code.