Network.AWS.Glue.Types.Action

Description

Documentation

data Action

Defines an action to be initiated by a trigger.

See: newAction smart constructor.

Constructors

Action'

Fields

securityConfiguration :: Maybe Text

The name of the SecurityConfiguration structure to be used with this action.

timeout :: Maybe Natural

The JobRun timeout in minutes. This is the maximum time that a job run can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours). This overrides the timeout value set in the parent job.

crawlerName :: Maybe Text

The name of the crawler to be used with this action.

notificationProperty :: Maybe NotificationProperty

Specifies configuration properties of a job run notification.

arguments :: Maybe (HashMap Text Text)

The job arguments used when this trigger fires. For this job run, they replace the default arguments set in the job definition itself.

You can specify arguments here that your own job-execution script consumes, as well as arguments that Glue itself consumes.

For information about how to specify and consume your own Job arguments, see the Calling Glue APIs in Python topic in the developer guide.

For information about the key-value pairs that Glue consumes to set up your job, see the Special Parameters Used by Glue topic in the developer guide.

jobName :: Maybe Text

The name of a job to be run.

newAction :: Action

Create a value of Action with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityConfiguration:Action', action_securityConfiguration - The name of the SecurityConfiguration structure to be used with this action.

$sel:timeout:Action', action_timeout - The JobRun timeout in minutes. This is the maximum time that a job run can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours). This overrides the timeout value set in the parent job.

$sel:crawlerName:Action', action_crawlerName - The name of the crawler to be used with this action.

$sel:notificationProperty:Action', action_notificationProperty - Specifies configuration properties of a job run notification.

$sel:arguments:Action', action_arguments - The job arguments used when this trigger fires. For this job run, they replace the default arguments set in the job definition itself.

You can specify arguments here that your own job-execution script consumes, as well as arguments that Glue itself consumes.

For information about how to specify and consume your own Job arguments, see the Calling Glue APIs in Python topic in the developer guide.

For information about the key-value pairs that Glue consumes to set up your job, see the Special Parameters Used by Glue topic in the developer guide.

$sel:jobName:Action', action_jobName - The name of a job to be run.

action_securityConfiguration :: Lens' Action (Maybe Text)

The name of the SecurityConfiguration structure to be used with this action.

action_timeout :: Lens' Action (Maybe Natural)

The JobRun timeout in minutes. This is the maximum time that a job run can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours). This overrides the timeout value set in the parent job.

action_crawlerName :: Lens' Action (Maybe Text)

The name of the crawler to be used with this action.

action_notificationProperty :: Lens' Action (Maybe NotificationProperty)

Specifies configuration properties of a job run notification.

action_arguments :: Lens' Action (Maybe (HashMap Text Text))

The job arguments used when this trigger fires. For this job run, they replace the default arguments set in the job definition itself.

You can specify arguments here that your own job-execution script consumes, as well as arguments that Glue itself consumes.

For information about how to specify and consume your own Job arguments, see the Calling Glue APIs in Python topic in the developer guide.

For information about the key-value pairs that Glue consumes to set up your job, see the Special Parameters Used by Glue topic in the developer guide.

action_jobName :: Lens' Action (Maybe Text)

The name of a job to be run.