Network.AWS.Glue.Types.BlueprintRun

Description

Documentation

data BlueprintRun

The details of a blueprint run.

See: newBlueprintRun smart constructor.

Constructors

BlueprintRun'

Fields

runId :: Maybe Text

The run ID for this blueprint run.

workflowName :: Maybe Text

The name of a workflow that is created as a result of a successful blueprint run. If a blueprint run has an error, there will not be a workflow created.

roleArn :: Maybe Text

The role ARN. This role will be assumed by the Glue service and will be used to create the workflow and other entities of a workflow.

blueprintName :: Maybe Text

The name of the blueprint.

state :: Maybe BlueprintRunState

The state of the blueprint run. Possible values are:

Running — The blueprint run is in progress.

Succeeded — The blueprint run completed successfully.

Failed — The blueprint run failed and rollback is complete.

Rolling Back — The blueprint run failed and rollback is in progress.

completedOn :: Maybe POSIX

The date and time that the blueprint run completed.

errorMessage :: Maybe Text

Indicates any errors that are seen while running the blueprint.

startedOn :: Maybe POSIX

The date and time that the blueprint run started.

rollbackErrorMessage :: Maybe Text

If there are any errors while creating the entities of a workflow, we try to roll back the created entities until that point and delete them. This attribute indicates the errors seen while trying to delete the entities that are created.

parameters :: Maybe Text

The blueprint parameters as a string. You will have to provide a value for each key that is required from the parameter spec that is defined in the Blueprint$ParameterSpec.

newBlueprintRun :: BlueprintRun

Create a value of BlueprintRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:runId:BlueprintRun', blueprintRun_runId - The run ID for this blueprint run.

$sel:workflowName:BlueprintRun', blueprintRun_workflowName - The name of a workflow that is created as a result of a successful blueprint run. If a blueprint run has an error, there will not be a workflow created.

$sel:roleArn:BlueprintRun', blueprintRun_roleArn - The role ARN. This role will be assumed by the Glue service and will be used to create the workflow and other entities of a workflow.

$sel:blueprintName:BlueprintRun', blueprintRun_blueprintName - The name of the blueprint.

$sel:state:BlueprintRun', blueprintRun_state - The state of the blueprint run. Possible values are:

Running — The blueprint run is in progress.

Succeeded — The blueprint run completed successfully.

Failed — The blueprint run failed and rollback is complete.

Rolling Back — The blueprint run failed and rollback is in progress.

$sel:completedOn:BlueprintRun', blueprintRun_completedOn - The date and time that the blueprint run completed.

$sel:errorMessage:BlueprintRun', blueprintRun_errorMessage - Indicates any errors that are seen while running the blueprint.

$sel:startedOn:BlueprintRun', blueprintRun_startedOn - The date and time that the blueprint run started.

$sel:rollbackErrorMessage:BlueprintRun', blueprintRun_rollbackErrorMessage - If there are any errors while creating the entities of a workflow, we try to roll back the created entities until that point and delete them. This attribute indicates the errors seen while trying to delete the entities that are created.

$sel:parameters:BlueprintRun', blueprintRun_parameters - The blueprint parameters as a string. You will have to provide a value for each key that is required from the parameter spec that is defined in the Blueprint$ParameterSpec.

blueprintRun_runId :: Lens' BlueprintRun (Maybe Text)

The run ID for this blueprint run.

blueprintRun_workflowName :: Lens' BlueprintRun (Maybe Text)

The name of a workflow that is created as a result of a successful blueprint run. If a blueprint run has an error, there will not be a workflow created.

blueprintRun_roleArn :: Lens' BlueprintRun (Maybe Text)

The role ARN. This role will be assumed by the Glue service and will be used to create the workflow and other entities of a workflow.

blueprintRun_blueprintName :: Lens' BlueprintRun (Maybe Text)

The name of the blueprint.

blueprintRun_state :: Lens' BlueprintRun (Maybe BlueprintRunState)

The state of the blueprint run. Possible values are:

Running — The blueprint run is in progress.

Succeeded — The blueprint run completed successfully.

Failed — The blueprint run failed and rollback is complete.

Rolling Back — The blueprint run failed and rollback is in progress.

blueprintRun_completedOn :: Lens' BlueprintRun (Maybe UTCTime)

The date and time that the blueprint run completed.

blueprintRun_errorMessage :: Lens' BlueprintRun (Maybe Text)

Indicates any errors that are seen while running the blueprint.

blueprintRun_startedOn :: Lens' BlueprintRun (Maybe UTCTime)

The date and time that the blueprint run started.

blueprintRun_rollbackErrorMessage :: Lens' BlueprintRun (Maybe Text)

If there are any errors while creating the entities of a workflow, we try to roll back the created entities until that point and delete them. This attribute indicates the errors seen while trying to delete the entities that are created.

blueprintRun_parameters :: Lens' BlueprintRun (Maybe Text)

The blueprint parameters as a string. You will have to provide a value for each key that is required from the parameter spec that is defined in the Blueprint$ParameterSpec.