Network.AWS.Glue.Types.ColumnError

Description

Documentation

data ColumnError

Encapsulates a column name that failed and the reason for failure.

See: newColumnError smart constructor.

Constructors

ColumnError'

Fields

columnName :: Maybe Text

The name of the column that failed.

error :: Maybe ErrorDetail

An error message with the reason for the failure of an operation.

newColumnError :: ColumnError

Create a value of ColumnError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:columnName:ColumnError', columnError_columnName - The name of the column that failed.

$sel:error:ColumnError', columnError_error - An error message with the reason for the failure of an operation.

columnError_columnName :: Lens' ColumnError (Maybe Text)

The name of the column that failed.

columnError_error :: Lens' ColumnError (Maybe ErrorDetail)

An error message with the reason for the failure of an operation.