Network.AWS.Glue.Types.ColumnStatisticsData

Description

Documentation

data ColumnStatisticsData

Contains the individual types of column statistics data. Only one data object should be set and indicated by the Type attribute.

See: newColumnStatisticsData smart constructor.

Constructors

ColumnStatisticsData'

Fields

dateColumnStatisticsData :: Maybe DateColumnStatisticsData

Date column statistics data.

binaryColumnStatisticsData :: Maybe BinaryColumnStatisticsData

Binary column statistics data.

booleanColumnStatisticsData :: Maybe BooleanColumnStatisticsData

Boolean column statistics data.

longColumnStatisticsData :: Maybe LongColumnStatisticsData

Long column statistics data.

stringColumnStatisticsData :: Maybe StringColumnStatisticsData

String column statistics data.

doubleColumnStatisticsData :: Maybe DoubleColumnStatisticsData

Double column statistics data.

decimalColumnStatisticsData :: Maybe DecimalColumnStatisticsData

Decimal column statistics data.

type' :: ColumnStatisticsType

The type of column statistics data.

newColumnStatisticsData

Arguments

:: ColumnStatisticsType
$sel:type':ColumnStatisticsData'

-> ColumnStatisticsData

Create a value of ColumnStatisticsData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dateColumnStatisticsData:ColumnStatisticsData', columnStatisticsData_dateColumnStatisticsData - Date column statistics data.

$sel:binaryColumnStatisticsData:ColumnStatisticsData', columnStatisticsData_binaryColumnStatisticsData - Binary column statistics data.

$sel:booleanColumnStatisticsData:ColumnStatisticsData', columnStatisticsData_booleanColumnStatisticsData - Boolean column statistics data.

$sel:longColumnStatisticsData:ColumnStatisticsData', columnStatisticsData_longColumnStatisticsData - Long column statistics data.

$sel:stringColumnStatisticsData:ColumnStatisticsData', columnStatisticsData_stringColumnStatisticsData - String column statistics data.

$sel:doubleColumnStatisticsData:ColumnStatisticsData', columnStatisticsData_doubleColumnStatisticsData - Double column statistics data.

$sel:decimalColumnStatisticsData:ColumnStatisticsData', columnStatisticsData_decimalColumnStatisticsData - Decimal column statistics data.

$sel:type':ColumnStatisticsData', columnStatisticsData_type - The type of column statistics data.

columnStatisticsData_dateColumnStatisticsData :: Lens' ColumnStatisticsData (Maybe DateColumnStatisticsData)

Date column statistics data.

columnStatisticsData_binaryColumnStatisticsData :: Lens' ColumnStatisticsData (Maybe BinaryColumnStatisticsData)

Binary column statistics data.

columnStatisticsData_booleanColumnStatisticsData :: Lens' ColumnStatisticsData (Maybe BooleanColumnStatisticsData)

Boolean column statistics data.

columnStatisticsData_longColumnStatisticsData :: Lens' ColumnStatisticsData (Maybe LongColumnStatisticsData)

Long column statistics data.

columnStatisticsData_stringColumnStatisticsData :: Lens' ColumnStatisticsData (Maybe StringColumnStatisticsData)

String column statistics data.

columnStatisticsData_doubleColumnStatisticsData :: Lens' ColumnStatisticsData (Maybe DoubleColumnStatisticsData)

Double column statistics data.

columnStatisticsData_decimalColumnStatisticsData :: Lens' ColumnStatisticsData (Maybe DecimalColumnStatisticsData)

Decimal column statistics data.

columnStatisticsData_type :: Lens' ColumnStatisticsData ColumnStatisticsType

The type of column statistics data.