Network.AWS.Glue.Types.Condition

Description

Documentation

data Condition

Defines a condition under which a trigger fires.

See: newCondition smart constructor.

Constructors

Condition'

Fields

crawlState :: Maybe CrawlState

The state of the crawler to which this condition applies.

state :: Maybe JobRunState

The condition state. Currently, the only job states that a trigger can listen for are SUCCEEDED, STOPPED, FAILED, and TIMEOUT. The only crawler states that a trigger can listen for are SUCCEEDED, FAILED, and CANCELLED.

jobName :: Maybe Text

The name of the job whose JobRuns this condition applies to, and on which this trigger waits.

logicalOperator :: Maybe LogicalOperator

A logical operator.

crawlerName :: Maybe Text

The name of the crawler to which this condition applies.

newCondition :: Condition

Create a value of Condition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:crawlState:Condition', condition_crawlState - The state of the crawler to which this condition applies.

$sel:state:Condition', condition_state - The condition state. Currently, the only job states that a trigger can listen for are SUCCEEDED, STOPPED, FAILED, and TIMEOUT. The only crawler states that a trigger can listen for are SUCCEEDED, FAILED, and CANCELLED.

$sel:jobName:Condition', condition_jobName - The name of the job whose JobRuns this condition applies to, and on which this trigger waits.

$sel:logicalOperator:Condition', condition_logicalOperator - A logical operator.

$sel:crawlerName:Condition', condition_crawlerName - The name of the crawler to which this condition applies.

condition_crawlState :: Lens' Condition (Maybe CrawlState)

The state of the crawler to which this condition applies.

condition_state :: Lens' Condition (Maybe JobRunState)

The condition state. Currently, the only job states that a trigger can listen for are SUCCEEDED, STOPPED, FAILED, and TIMEOUT. The only crawler states that a trigger can listen for are SUCCEEDED, FAILED, and CANCELLED.

condition_jobName :: Lens' Condition (Maybe Text)

The name of the job whose JobRuns this condition applies to, and on which this trigger waits.

condition_logicalOperator :: Lens' Condition (Maybe LogicalOperator)

A logical operator.

condition_crawlerName :: Lens' Condition (Maybe Text)

The name of the crawler to which this condition applies.