Network.AWS.Glue.Types.ConfusionMatrix

Description

Documentation

data ConfusionMatrix

The confusion matrix shows you what your transform is predicting accurately and what types of errors it is making.

For more information, see Confusion matrix in Wikipedia.

See: newConfusionMatrix smart constructor.

Constructors

ConfusionMatrix'

Fields

numFalsePositives :: Maybe Integer

The number of nonmatches in the data that the transform incorrectly classified as a match, in the confusion matrix for your transform.

numTrueNegatives :: Maybe Integer

The number of nonmatches in the data that the transform correctly rejected, in the confusion matrix for your transform.

numFalseNegatives :: Maybe Integer

The number of matches in the data that the transform didn't find, in the confusion matrix for your transform.

numTruePositives :: Maybe Integer

The number of matches in the data that the transform correctly found, in the confusion matrix for your transform.

newConfusionMatrix :: ConfusionMatrix

Create a value of ConfusionMatrix with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:numFalsePositives:ConfusionMatrix', confusionMatrix_numFalsePositives - The number of nonmatches in the data that the transform incorrectly classified as a match, in the confusion matrix for your transform.

$sel:numTrueNegatives:ConfusionMatrix', confusionMatrix_numTrueNegatives - The number of nonmatches in the data that the transform correctly rejected, in the confusion matrix for your transform.

$sel:numFalseNegatives:ConfusionMatrix', confusionMatrix_numFalseNegatives - The number of matches in the data that the transform didn't find, in the confusion matrix for your transform.

$sel:numTruePositives:ConfusionMatrix', confusionMatrix_numTruePositives - The number of matches in the data that the transform correctly found, in the confusion matrix for your transform.

confusionMatrix_numFalsePositives :: Lens' ConfusionMatrix (Maybe Integer)

The number of nonmatches in the data that the transform incorrectly classified as a match, in the confusion matrix for your transform.

confusionMatrix_numTrueNegatives :: Lens' ConfusionMatrix (Maybe Integer)

The number of nonmatches in the data that the transform correctly rejected, in the confusion matrix for your transform.

confusionMatrix_numFalseNegatives :: Lens' ConfusionMatrix (Maybe Integer)

The number of matches in the data that the transform didn't find, in the confusion matrix for your transform.

confusionMatrix_numTruePositives :: Lens' ConfusionMatrix (Maybe Integer)

The number of matches in the data that the transform correctly found, in the confusion matrix for your transform.