Network.AWS.Glue.Types.Connection

Description

Documentation

data Connection

Defines a connection to a data source.

See: newConnection smart constructor.

Constructors

Connection'

Fields

connectionProperties :: Maybe (HashMap ConnectionPropertyKey Text)

These key-value pairs define parameters for the connection:

HOST - The host URI: either the fully qualified domain name (FQDN) or the IPv4 address of the database host.

PORT - The port number, between 1024 and 65535, of the port on which the database host is listening for database connections.

USER_NAME - The name under which to log in to the database. The value string for USER_NAME is "USERNAME".

PASSWORD - A password, if one is used, for the user name.

ENCRYPTED_PASSWORD - When you enable connection password protection by setting ConnectionPasswordEncryption in the Data Catalog encryption settings, this field stores the encrypted password.

JDBC_DRIVER_JAR_URI - The Amazon Simple Storage Service (Amazon S3) path of the JAR file that contains the JDBC driver to use.

JDBC_DRIVER_CLASS_NAME - The class name of the JDBC driver to use.

JDBC_ENGINE - The name of the JDBC engine to use.

JDBC_ENGINE_VERSION - The version of the JDBC engine to use.

CONFIG_FILES - (Reserved for future use.)

INSTANCE_ID - The instance ID to use.

JDBC_CONNECTION_URL - The URL for connecting to a JDBC data source.

JDBC_ENFORCE_SSL - A Boolean string (true, false) specifying whether Secure Sockets Layer (SSL) with hostname matching is enforced for the JDBC connection on the client. The default is false.

CUSTOM_JDBC_CERT - An Amazon S3 location specifying the customer's root certificate. Glue uses this root certificate to validate the customer’s certificate when connecting to the customer database. Glue only handles X.509 certificates. The certificate provided must be DER-encoded and supplied in Base64 encoding PEM format.

SKIP_CUSTOM_JDBC_CERT_VALIDATION - By default, this is false. Glue validates the Signature algorithm and Subject Public Key Algorithm for the customer certificate. The only permitted algorithms for the Signature algorithm are SHA256withRSA, SHA384withRSA or SHA512withRSA. For the Subject Public Key Algorithm, the key length must be at least 2048. You can set the value of this property to true to skip Glue’s validation of the customer certificate.

CUSTOM_JDBC_CERT_STRING - A custom JDBC certificate string which is used for domain match or distinguished name match to prevent a man-in-the-middle attack. In Oracle database, this is used as the SSL_SERVER_CERT_DN; in Microsoft SQL Server, this is used as the hostNameInCertificate.

CONNECTION_URL - The URL for connecting to a general (non-JDBC) data source.

KAFKA_BOOTSTRAP_SERVERS - A comma-separated list of host and port pairs that are the addresses of the Apache Kafka brokers in a Kafka cluster to which a Kafka client will connect to and bootstrap itself.

KAFKA_SSL_ENABLED - Whether to enable or disable SSL on an Apache Kafka connection. Default value is "true".

KAFKA_CUSTOM_CERT - The Amazon S3 URL for the private CA cert file (.pem format). The default is an empty string.

KAFKA_SKIP_CUSTOM_CERT_VALIDATION - Whether to skip the validation of the CA cert file or not. Glue validates for three algorithms: SHA256withRSA, SHA384withRSA and SHA512withRSA. Default value is "false".

SECRET_ID - The secret ID used for the secret manager of credentials.

CONNECTOR_URL - The connector URL for a MARKETPLACE or CUSTOM connection.

CONNECTOR_TYPE - The connector type for a MARKETPLACE or CUSTOM connection.

CONNECTOR_CLASS_NAME - The connector class name for a MARKETPLACE or CUSTOM connection.

KAFKA_CLIENT_KEYSTORE - The Amazon S3 location of the client keystore file for Kafka client side authentication (Optional).

KAFKA_CLIENT_KEYSTORE_PASSWORD - The password to access the provided keystore (Optional).

KAFKA_CLIENT_KEY_PASSWORD - A keystore can consist of multiple keys, so this is the password to access the client key to be used with the Kafka server side key (Optional).

ENCRYPTED_KAFKA_CLIENT_KEYSTORE_PASSWORD - The encrypted version of the Kafka client keystore password (if the user has the Glue encrypt passwords setting selected).

ENCRYPTED_KAFKA_CLIENT_KEY_PASSWORD - The encrypted version of the Kafka client key password (if the user has the Glue encrypt passwords setting selected).

creationTime :: Maybe POSIX

The time that this connection definition was created.

connectionType :: Maybe ConnectionType

The type of the connection. Currently, SFTP is not supported.

name :: Maybe Text

The name of the connection definition.

physicalConnectionRequirements :: Maybe PhysicalConnectionRequirements

A map of physical connection requirements, such as virtual private cloud (VPC) and SecurityGroup, that are needed to make this connection successfully.

lastUpdatedBy :: Maybe Text

The user, group, or role that last updated this connection definition.

description :: Maybe Text

The description of the connection.

matchCriteria :: Maybe [Text]

A list of criteria that can be used in selecting this connection.

lastUpdatedTime :: Maybe POSIX

The last time that this connection definition was updated.

newConnection :: Connection

Create a value of Connection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectionProperties:Connection', connection_connectionProperties - These key-value pairs define parameters for the connection:

HOST - The host URI: either the fully qualified domain name (FQDN) or the IPv4 address of the database host.

PORT - The port number, between 1024 and 65535, of the port on which the database host is listening for database connections.

USER_NAME - The name under which to log in to the database. The value string for USER_NAME is "USERNAME".

PASSWORD - A password, if one is used, for the user name.

ENCRYPTED_PASSWORD - When you enable connection password protection by setting ConnectionPasswordEncryption in the Data Catalog encryption settings, this field stores the encrypted password.

JDBC_DRIVER_JAR_URI - The Amazon Simple Storage Service (Amazon S3) path of the JAR file that contains the JDBC driver to use.

JDBC_DRIVER_CLASS_NAME - The class name of the JDBC driver to use.

JDBC_ENGINE - The name of the JDBC engine to use.

JDBC_ENGINE_VERSION - The version of the JDBC engine to use.

CONFIG_FILES - (Reserved for future use.)

INSTANCE_ID - The instance ID to use.

JDBC_CONNECTION_URL - The URL for connecting to a JDBC data source.

JDBC_ENFORCE_SSL - A Boolean string (true, false) specifying whether Secure Sockets Layer (SSL) with hostname matching is enforced for the JDBC connection on the client. The default is false.

CUSTOM_JDBC_CERT - An Amazon S3 location specifying the customer's root certificate. Glue uses this root certificate to validate the customer’s certificate when connecting to the customer database. Glue only handles X.509 certificates. The certificate provided must be DER-encoded and supplied in Base64 encoding PEM format.

SKIP_CUSTOM_JDBC_CERT_VALIDATION - By default, this is false. Glue validates the Signature algorithm and Subject Public Key Algorithm for the customer certificate. The only permitted algorithms for the Signature algorithm are SHA256withRSA, SHA384withRSA or SHA512withRSA. For the Subject Public Key Algorithm, the key length must be at least 2048. You can set the value of this property to true to skip Glue’s validation of the customer certificate.

CUSTOM_JDBC_CERT_STRING - A custom JDBC certificate string which is used for domain match or distinguished name match to prevent a man-in-the-middle attack. In Oracle database, this is used as the SSL_SERVER_CERT_DN; in Microsoft SQL Server, this is used as the hostNameInCertificate.

CONNECTION_URL - The URL for connecting to a general (non-JDBC) data source.

KAFKA_BOOTSTRAP_SERVERS - A comma-separated list of host and port pairs that are the addresses of the Apache Kafka brokers in a Kafka cluster to which a Kafka client will connect to and bootstrap itself.

KAFKA_SSL_ENABLED - Whether to enable or disable SSL on an Apache Kafka connection. Default value is "true".

KAFKA_CUSTOM_CERT - The Amazon S3 URL for the private CA cert file (.pem format). The default is an empty string.

KAFKA_SKIP_CUSTOM_CERT_VALIDATION - Whether to skip the validation of the CA cert file or not. Glue validates for three algorithms: SHA256withRSA, SHA384withRSA and SHA512withRSA. Default value is "false".

SECRET_ID - The secret ID used for the secret manager of credentials.

CONNECTOR_URL - The connector URL for a MARKETPLACE or CUSTOM connection.

CONNECTOR_TYPE - The connector type for a MARKETPLACE or CUSTOM connection.

CONNECTOR_CLASS_NAME - The connector class name for a MARKETPLACE or CUSTOM connection.

KAFKA_CLIENT_KEYSTORE - The Amazon S3 location of the client keystore file for Kafka client side authentication (Optional).

KAFKA_CLIENT_KEYSTORE_PASSWORD - The password to access the provided keystore (Optional).

KAFKA_CLIENT_KEY_PASSWORD - A keystore can consist of multiple keys, so this is the password to access the client key to be used with the Kafka server side key (Optional).

ENCRYPTED_KAFKA_CLIENT_KEYSTORE_PASSWORD - The encrypted version of the Kafka client keystore password (if the user has the Glue encrypt passwords setting selected).

ENCRYPTED_KAFKA_CLIENT_KEY_PASSWORD - The encrypted version of the Kafka client key password (if the user has the Glue encrypt passwords setting selected).

$sel:creationTime:Connection', connection_creationTime - The time that this connection definition was created.

$sel:connectionType:Connection', connection_connectionType - The type of the connection. Currently, SFTP is not supported.

$sel:name:Connection', connection_name - The name of the connection definition.

$sel:physicalConnectionRequirements:Connection', connection_physicalConnectionRequirements - A map of physical connection requirements, such as virtual private cloud (VPC) and SecurityGroup, that are needed to make this connection successfully.

$sel:lastUpdatedBy:Connection', connection_lastUpdatedBy - The user, group, or role that last updated this connection definition.

$sel:description:Connection', connection_description - The description of the connection.

$sel:matchCriteria:Connection', connection_matchCriteria - A list of criteria that can be used in selecting this connection.

$sel:lastUpdatedTime:Connection', connection_lastUpdatedTime - The last time that this connection definition was updated.

connection_connectionProperties :: Lens' Connection (Maybe (HashMap ConnectionPropertyKey Text))

These key-value pairs define parameters for the connection:

HOST - The host URI: either the fully qualified domain name (FQDN) or the IPv4 address of the database host.

PORT - The port number, between 1024 and 65535, of the port on which the database host is listening for database connections.

USER_NAME - The name under which to log in to the database. The value string for USER_NAME is "USERNAME".

PASSWORD - A password, if one is used, for the user name.

ENCRYPTED_PASSWORD - When you enable connection password protection by setting ConnectionPasswordEncryption in the Data Catalog encryption settings, this field stores the encrypted password.

JDBC_DRIVER_JAR_URI - The Amazon Simple Storage Service (Amazon S3) path of the JAR file that contains the JDBC driver to use.

JDBC_DRIVER_CLASS_NAME - The class name of the JDBC driver to use.

JDBC_ENGINE - The name of the JDBC engine to use.

JDBC_ENGINE_VERSION - The version of the JDBC engine to use.

CONFIG_FILES - (Reserved for future use.)

INSTANCE_ID - The instance ID to use.

JDBC_CONNECTION_URL - The URL for connecting to a JDBC data source.

JDBC_ENFORCE_SSL - A Boolean string (true, false) specifying whether Secure Sockets Layer (SSL) with hostname matching is enforced for the JDBC connection on the client. The default is false.

CUSTOM_JDBC_CERT - An Amazon S3 location specifying the customer's root certificate. Glue uses this root certificate to validate the customer’s certificate when connecting to the customer database. Glue only handles X.509 certificates. The certificate provided must be DER-encoded and supplied in Base64 encoding PEM format.

SKIP_CUSTOM_JDBC_CERT_VALIDATION - By default, this is false. Glue validates the Signature algorithm and Subject Public Key Algorithm for the customer certificate. The only permitted algorithms for the Signature algorithm are SHA256withRSA, SHA384withRSA or SHA512withRSA. For the Subject Public Key Algorithm, the key length must be at least 2048. You can set the value of this property to true to skip Glue’s validation of the customer certificate.

CUSTOM_JDBC_CERT_STRING - A custom JDBC certificate string which is used for domain match or distinguished name match to prevent a man-in-the-middle attack. In Oracle database, this is used as the SSL_SERVER_CERT_DN; in Microsoft SQL Server, this is used as the hostNameInCertificate.

CONNECTION_URL - The URL for connecting to a general (non-JDBC) data source.

KAFKA_BOOTSTRAP_SERVERS - A comma-separated list of host and port pairs that are the addresses of the Apache Kafka brokers in a Kafka cluster to which a Kafka client will connect to and bootstrap itself.

KAFKA_SSL_ENABLED - Whether to enable or disable SSL on an Apache Kafka connection. Default value is "true".

KAFKA_CUSTOM_CERT - The Amazon S3 URL for the private CA cert file (.pem format). The default is an empty string.

KAFKA_SKIP_CUSTOM_CERT_VALIDATION - Whether to skip the validation of the CA cert file or not. Glue validates for three algorithms: SHA256withRSA, SHA384withRSA and SHA512withRSA. Default value is "false".

SECRET_ID - The secret ID used for the secret manager of credentials.

CONNECTOR_URL - The connector URL for a MARKETPLACE or CUSTOM connection.

CONNECTOR_TYPE - The connector type for a MARKETPLACE or CUSTOM connection.

CONNECTOR_CLASS_NAME - The connector class name for a MARKETPLACE or CUSTOM connection.

KAFKA_CLIENT_KEYSTORE - The Amazon S3 location of the client keystore file for Kafka client side authentication (Optional).

KAFKA_CLIENT_KEYSTORE_PASSWORD - The password to access the provided keystore (Optional).

KAFKA_CLIENT_KEY_PASSWORD - A keystore can consist of multiple keys, so this is the password to access the client key to be used with the Kafka server side key (Optional).

ENCRYPTED_KAFKA_CLIENT_KEYSTORE_PASSWORD - The encrypted version of the Kafka client keystore password (if the user has the Glue encrypt passwords setting selected).

ENCRYPTED_KAFKA_CLIENT_KEY_PASSWORD - The encrypted version of the Kafka client key password (if the user has the Glue encrypt passwords setting selected).

connection_creationTime :: Lens' Connection (Maybe UTCTime)

The time that this connection definition was created.

connection_connectionType :: Lens' Connection (Maybe ConnectionType)

The type of the connection. Currently, SFTP is not supported.

connection_name :: Lens' Connection (Maybe Text)

The name of the connection definition.

connection_physicalConnectionRequirements :: Lens' Connection (Maybe PhysicalConnectionRequirements)

A map of physical connection requirements, such as virtual private cloud (VPC) and SecurityGroup, that are needed to make this connection successfully.

connection_lastUpdatedBy :: Lens' Connection (Maybe Text)

The user, group, or role that last updated this connection definition.

connection_description :: Lens' Connection (Maybe Text)

The description of the connection.

connection_matchCriteria :: Lens' Connection (Maybe [Text])

A list of criteria that can be used in selecting this connection.

connection_lastUpdatedTime :: Lens' Connection (Maybe UTCTime)

The last time that this connection definition was updated.