Network.AWS.Glue.Types.ConnectionInput

Description

Documentation

data ConnectionInput

A structure that is used to specify a connection to create or update.

See: newConnectionInput smart constructor.

Constructors

ConnectionInput'

Fields

matchCriteria :: Maybe [Text]

A list of criteria that can be used in selecting this connection.

physicalConnectionRequirements :: Maybe PhysicalConnectionRequirements

A map of physical connection requirements, such as virtual private cloud (VPC) and SecurityGroup, that are needed to successfully make this connection.

description :: Maybe Text

The description of the connection.

name :: Text

The name of the connection.

connectionType :: ConnectionType

The type of the connection. Currently, these types are supported:

JDBC - Designates a connection to a database through Java Database Connectivity (JDBC).

KAFKA - Designates a connection to an Apache Kafka streaming platform.

MONGODB - Designates a connection to a MongoDB document database.

NETWORK - Designates a network connection to a data source within an Amazon Virtual Private Cloud environment (Amazon VPC).

MARKETPLACE - Uses configuration settings contained in a connector purchased from Amazon Web Services Marketplace to read from and write to data stores that are not natively supported by Glue.

CUSTOM - Uses configuration settings contained in a custom connector to read from and write to data stores that are not natively supported by Glue.

SFTP is not supported.

connectionProperties :: HashMap ConnectionPropertyKey Text

These key-value pairs define parameters for the connection.

newConnectionInput

Arguments

:: Text
$sel:name:ConnectionInput'

-> ConnectionType
$sel:connectionType:ConnectionInput'

-> ConnectionInput

Create a value of ConnectionInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:matchCriteria:ConnectionInput', connectionInput_matchCriteria - A list of criteria that can be used in selecting this connection.

$sel:physicalConnectionRequirements:ConnectionInput', connectionInput_physicalConnectionRequirements - A map of physical connection requirements, such as virtual private cloud (VPC) and SecurityGroup, that are needed to successfully make this connection.

$sel:description:ConnectionInput', connectionInput_description - The description of the connection.

$sel:name:ConnectionInput', connectionInput_name - The name of the connection.

$sel:connectionType:ConnectionInput', connectionInput_connectionType - The type of the connection. Currently, these types are supported:

JDBC - Designates a connection to a database through Java Database Connectivity (JDBC).

KAFKA - Designates a connection to an Apache Kafka streaming platform.

MONGODB - Designates a connection to a MongoDB document database.

NETWORK - Designates a network connection to a data source within an Amazon Virtual Private Cloud environment (Amazon VPC).

MARKETPLACE - Uses configuration settings contained in a connector purchased from Amazon Web Services Marketplace to read from and write to data stores that are not natively supported by Glue.

CUSTOM - Uses configuration settings contained in a custom connector to read from and write to data stores that are not natively supported by Glue.

SFTP is not supported.

$sel:connectionProperties:ConnectionInput', connectionInput_connectionProperties - These key-value pairs define parameters for the connection.

connectionInput_matchCriteria :: Lens' ConnectionInput (Maybe [Text])

A list of criteria that can be used in selecting this connection.

connectionInput_physicalConnectionRequirements :: Lens' ConnectionInput (Maybe PhysicalConnectionRequirements)

A map of physical connection requirements, such as virtual private cloud (VPC) and SecurityGroup, that are needed to successfully make this connection.

connectionInput_description :: Lens' ConnectionInput (Maybe Text)

The description of the connection.

connectionInput_name :: Lens' ConnectionInput Text

The name of the connection.

connectionInput_connectionType :: Lens' ConnectionInput ConnectionType

The type of the connection. Currently, these types are supported:

JDBC - Designates a connection to a database through Java Database Connectivity (JDBC).

KAFKA - Designates a connection to an Apache Kafka streaming platform.

MONGODB - Designates a connection to a MongoDB document database.

NETWORK - Designates a network connection to a data source within an Amazon Virtual Private Cloud environment (Amazon VPC).

MARKETPLACE - Uses configuration settings contained in a connector purchased from Amazon Web Services Marketplace to read from and write to data stores that are not natively supported by Glue.

CUSTOM - Uses configuration settings contained in a custom connector to read from and write to data stores that are not natively supported by Glue.

SFTP is not supported.

connectionInput_connectionProperties :: Lens' ConnectionInput (HashMap ConnectionPropertyKey Text)

These key-value pairs define parameters for the connection.