Network.AWS.Glue.Types.Crawl

Description

Documentation

data Crawl

The details of a crawl in the workflow.

See: newCrawl smart constructor.

Constructors

Crawl'

Fields

logGroup :: Maybe Text

The log group associated with the crawl.

state :: Maybe CrawlState

The state of the crawler.

completedOn :: Maybe POSIX

The date and time on which the crawl completed.

errorMessage :: Maybe Text

The error message associated with the crawl.

startedOn :: Maybe POSIX

The date and time on which the crawl started.

logStream :: Maybe Text

The log stream associated with the crawl.

newCrawl :: Crawl

Create a value of Crawl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logGroup:Crawl', crawl_logGroup - The log group associated with the crawl.

$sel:state:Crawl', crawl_state - The state of the crawler.

$sel:completedOn:Crawl', crawl_completedOn - The date and time on which the crawl completed.

$sel:errorMessage:Crawl', crawl_errorMessage - The error message associated with the crawl.

$sel:startedOn:Crawl', crawl_startedOn - The date and time on which the crawl started.

$sel:logStream:Crawl', crawl_logStream - The log stream associated with the crawl.

crawl_logGroup :: Lens' Crawl (Maybe Text)

The log group associated with the crawl.

crawl_state :: Lens' Crawl (Maybe CrawlState)

The state of the crawler.

crawl_completedOn :: Lens' Crawl (Maybe UTCTime)

The date and time on which the crawl completed.

crawl_errorMessage :: Lens' Crawl (Maybe Text)

The error message associated with the crawl.

crawl_startedOn :: Lens' Crawl (Maybe UTCTime)

The date and time on which the crawl started.

crawl_logStream :: Lens' Crawl (Maybe Text)

The log stream associated with the crawl.