Network.AWS.Glue.Types.Crawler

Description

Documentation

data Crawler

Specifies a crawler program that examines a data source and uses classifiers to try to determine its schema. If successful, the crawler records metadata concerning the data source in the Glue Data Catalog.

See: newCrawler smart constructor.

Constructors

Crawler'

Fields

schemaChangePolicy :: Maybe SchemaChangePolicy

The policy that specifies update and delete behaviors for the crawler.

creationTime :: Maybe POSIX

The time that the crawler was created.

recrawlPolicy :: Maybe RecrawlPolicy

A policy that specifies whether to crawl the entire dataset again, or to crawl only folders that were added since the last crawler run.

classifiers :: Maybe [Text]

A list of UTF-8 strings that specify the custom classifiers that are associated with the crawler.

configuration :: Maybe Text

Crawler configuration information. This versioned JSON string allows users to specify aspects of a crawler's behavior. For more information, see Include and Exclude Patterns.

lineageConfiguration :: Maybe LineageConfiguration

A configuration that specifies whether data lineage is enabled for the crawler.

version :: Maybe Integer

The version of the crawler.

targets :: Maybe CrawlerTargets

A collection of targets to crawl.

name :: Maybe Text

The name of the crawler.

state :: Maybe CrawlerState

Indicates whether the crawler is running, or whether a run is pending.

lastUpdated :: Maybe POSIX

The time that the crawler was last updated.

crawlElapsedTime :: Maybe Integer

If the crawler is running, contains the total time elapsed since the last crawl began.

role' :: Maybe Text

The Amazon Resource Name (ARN) of an IAM role that's used to access customer resources, such as Amazon Simple Storage Service (Amazon S3) data.

lastCrawl :: Maybe LastCrawlInfo

The status of the last crawl, and potentially error information if an error occurred.

description :: Maybe Text

A description of the crawler.

tablePrefix :: Maybe Text

The prefix added to the names of tables that are created.

schedule :: Maybe Schedule

For scheduled crawlers, the schedule when the crawler runs.

crawlerSecurityConfiguration :: Maybe Text

The name of the SecurityConfiguration structure to be used by this crawler.

databaseName :: Maybe Text

The name of the database in which the crawler's output is stored.

newCrawler :: Crawler

Create a value of Crawler with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaChangePolicy:Crawler', crawler_schemaChangePolicy - The policy that specifies update and delete behaviors for the crawler.

$sel:creationTime:Crawler', crawler_creationTime - The time that the crawler was created.

$sel:recrawlPolicy:Crawler', crawler_recrawlPolicy - A policy that specifies whether to crawl the entire dataset again, or to crawl only folders that were added since the last crawler run.

$sel:classifiers:Crawler', crawler_classifiers - A list of UTF-8 strings that specify the custom classifiers that are associated with the crawler.

$sel:configuration:Crawler', crawler_configuration - Crawler configuration information. This versioned JSON string allows users to specify aspects of a crawler's behavior. For more information, see Include and Exclude Patterns.

$sel:lineageConfiguration:Crawler', crawler_lineageConfiguration - A configuration that specifies whether data lineage is enabled for the crawler.

$sel:version:Crawler', crawler_version - The version of the crawler.

$sel:targets:Crawler', crawler_targets - A collection of targets to crawl.

$sel:name:Crawler', crawler_name - The name of the crawler.

$sel:state:Crawler', crawler_state - Indicates whether the crawler is running, or whether a run is pending.

$sel:lastUpdated:Crawler', crawler_lastUpdated - The time that the crawler was last updated.

$sel:crawlElapsedTime:Crawler', crawler_crawlElapsedTime - If the crawler is running, contains the total time elapsed since the last crawl began.

$sel:role':Crawler', crawler_role - The Amazon Resource Name (ARN) of an IAM role that's used to access customer resources, such as Amazon Simple Storage Service (Amazon S3) data.

$sel:lastCrawl:Crawler', crawler_lastCrawl - The status of the last crawl, and potentially error information if an error occurred.

$sel:description:Crawler', crawler_description - A description of the crawler.

$sel:tablePrefix:Crawler', crawler_tablePrefix - The prefix added to the names of tables that are created.

$sel:schedule:Crawler', crawler_schedule - For scheduled crawlers, the schedule when the crawler runs.

$sel:crawlerSecurityConfiguration:Crawler', crawler_crawlerSecurityConfiguration - The name of the SecurityConfiguration structure to be used by this crawler.

$sel:databaseName:Crawler', crawler_databaseName - The name of the database in which the crawler's output is stored.

crawler_schemaChangePolicy :: Lens' Crawler (Maybe SchemaChangePolicy)

The policy that specifies update and delete behaviors for the crawler.

crawler_creationTime :: Lens' Crawler (Maybe UTCTime)

The time that the crawler was created.

crawler_recrawlPolicy :: Lens' Crawler (Maybe RecrawlPolicy)

A policy that specifies whether to crawl the entire dataset again, or to crawl only folders that were added since the last crawler run.

crawler_classifiers :: Lens' Crawler (Maybe [Text])

A list of UTF-8 strings that specify the custom classifiers that are associated with the crawler.

crawler_configuration :: Lens' Crawler (Maybe Text)

Crawler configuration information. This versioned JSON string allows users to specify aspects of a crawler's behavior. For more information, see Include and Exclude Patterns.

crawler_lineageConfiguration :: Lens' Crawler (Maybe LineageConfiguration)

A configuration that specifies whether data lineage is enabled for the crawler.

crawler_version :: Lens' Crawler (Maybe Integer)

The version of the crawler.

crawler_targets :: Lens' Crawler (Maybe CrawlerTargets)

A collection of targets to crawl.

crawler_name :: Lens' Crawler (Maybe Text)

The name of the crawler.

crawler_state :: Lens' Crawler (Maybe CrawlerState)

Indicates whether the crawler is running, or whether a run is pending.

crawler_lastUpdated :: Lens' Crawler (Maybe UTCTime)

The time that the crawler was last updated.

crawler_crawlElapsedTime :: Lens' Crawler (Maybe Integer)

If the crawler is running, contains the total time elapsed since the last crawl began.

crawler_role :: Lens' Crawler (Maybe Text)

The Amazon Resource Name (ARN) of an IAM role that's used to access customer resources, such as Amazon Simple Storage Service (Amazon S3) data.

crawler_lastCrawl :: Lens' Crawler (Maybe LastCrawlInfo)

The status of the last crawl, and potentially error information if an error occurred.

crawler_description :: Lens' Crawler (Maybe Text)

A description of the crawler.

crawler_tablePrefix :: Lens' Crawler (Maybe Text)

The prefix added to the names of tables that are created.

crawler_schedule :: Lens' Crawler (Maybe Schedule)

For scheduled crawlers, the schedule when the crawler runs.

crawler_crawlerSecurityConfiguration :: Lens' Crawler (Maybe Text)

The name of the SecurityConfiguration structure to be used by this crawler.

crawler_databaseName :: Lens' Crawler (Maybe Text)

The name of the database in which the crawler's output is stored.