Network.AWS.Glue.Types.CrawlerMetrics

Description

Documentation

data CrawlerMetrics

Metrics for a specified crawler.

See: newCrawlerMetrics smart constructor.

Constructors

CrawlerMetrics'

Fields

lastRuntimeSeconds :: Maybe Double

The duration of the crawler's most recent run, in seconds.

tablesCreated :: Maybe Natural

The number of tables created by this crawler.

stillEstimating :: Maybe Bool

True if the crawler is still estimating how long it will take to complete this run.

medianRuntimeSeconds :: Maybe Double

The median duration of this crawler's runs, in seconds.

timeLeftSeconds :: Maybe Double

The estimated time left to complete a running crawl.

tablesDeleted :: Maybe Natural

The number of tables deleted by this crawler.

tablesUpdated :: Maybe Natural

The number of tables updated by this crawler.

crawlerName :: Maybe Text

The name of the crawler.

newCrawlerMetrics :: CrawlerMetrics

Create a value of CrawlerMetrics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastRuntimeSeconds:CrawlerMetrics', crawlerMetrics_lastRuntimeSeconds - The duration of the crawler's most recent run, in seconds.

$sel:tablesCreated:CrawlerMetrics', crawlerMetrics_tablesCreated - The number of tables created by this crawler.

$sel:stillEstimating:CrawlerMetrics', crawlerMetrics_stillEstimating - True if the crawler is still estimating how long it will take to complete this run.

$sel:medianRuntimeSeconds:CrawlerMetrics', crawlerMetrics_medianRuntimeSeconds - The median duration of this crawler's runs, in seconds.

$sel:timeLeftSeconds:CrawlerMetrics', crawlerMetrics_timeLeftSeconds - The estimated time left to complete a running crawl.

$sel:tablesDeleted:CrawlerMetrics', crawlerMetrics_tablesDeleted - The number of tables deleted by this crawler.

$sel:tablesUpdated:CrawlerMetrics', crawlerMetrics_tablesUpdated - The number of tables updated by this crawler.

$sel:crawlerName:CrawlerMetrics', crawlerMetrics_crawlerName - The name of the crawler.

crawlerMetrics_lastRuntimeSeconds :: Lens' CrawlerMetrics (Maybe Double)

The duration of the crawler's most recent run, in seconds.

crawlerMetrics_tablesCreated :: Lens' CrawlerMetrics (Maybe Natural)

The number of tables created by this crawler.

crawlerMetrics_stillEstimating :: Lens' CrawlerMetrics (Maybe Bool)

True if the crawler is still estimating how long it will take to complete this run.

crawlerMetrics_medianRuntimeSeconds :: Lens' CrawlerMetrics (Maybe Double)

The median duration of this crawler's runs, in seconds.

crawlerMetrics_timeLeftSeconds :: Lens' CrawlerMetrics (Maybe Double)

The estimated time left to complete a running crawl.

crawlerMetrics_tablesDeleted :: Lens' CrawlerMetrics (Maybe Natural)

The number of tables deleted by this crawler.

crawlerMetrics_tablesUpdated :: Lens' CrawlerMetrics (Maybe Natural)

The number of tables updated by this crawler.

crawlerMetrics_crawlerName :: Lens' CrawlerMetrics (Maybe Text)

The name of the crawler.