Network.AWS.Glue.Types.CreateCsvClassifierRequest

Description

Documentation

data CreateCsvClassifierRequest

Specifies a custom CSV classifier for CreateClassifier to create.

See: newCreateCsvClassifierRequest smart constructor.

Constructors

CreateCsvClassifierRequest'

Fields

quoteSymbol :: Maybe Text

A custom symbol to denote what combines content into a single column value. Must be different from the column delimiter.

containsHeader :: Maybe CsvHeaderOption

Indicates whether the CSV file contains a header.

disableValueTrimming :: Maybe Bool

Specifies not to trim values before identifying the type of column values. The default value is true.

header :: Maybe [Text]

A list of strings representing column names.

allowSingleColumn :: Maybe Bool

Enables the processing of files that contain only one column.

delimiter :: Maybe Text

A custom symbol to denote what separates each column entry in the row.

name :: Text

The name of the classifier.

newCreateCsvClassifierRequest

Arguments

:: Text
$sel:name:CreateCsvClassifierRequest'

-> CreateCsvClassifierRequest

Create a value of CreateCsvClassifierRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:quoteSymbol:CreateCsvClassifierRequest', createCsvClassifierRequest_quoteSymbol - A custom symbol to denote what combines content into a single column value. Must be different from the column delimiter.

$sel:containsHeader:CreateCsvClassifierRequest', createCsvClassifierRequest_containsHeader - Indicates whether the CSV file contains a header.

$sel:disableValueTrimming:CreateCsvClassifierRequest', createCsvClassifierRequest_disableValueTrimming - Specifies not to trim values before identifying the type of column values. The default value is true.

$sel:header:CreateCsvClassifierRequest', createCsvClassifierRequest_header - A list of strings representing column names.

$sel:allowSingleColumn:CreateCsvClassifierRequest', createCsvClassifierRequest_allowSingleColumn - Enables the processing of files that contain only one column.

$sel:delimiter:CreateCsvClassifierRequest', createCsvClassifierRequest_delimiter - A custom symbol to denote what separates each column entry in the row.

$sel:name:CreateCsvClassifierRequest', createCsvClassifierRequest_name - The name of the classifier.

createCsvClassifierRequest_quoteSymbol :: Lens' CreateCsvClassifierRequest (Maybe Text)

A custom symbol to denote what combines content into a single column value. Must be different from the column delimiter.

createCsvClassifierRequest_containsHeader :: Lens' CreateCsvClassifierRequest (Maybe CsvHeaderOption)

Indicates whether the CSV file contains a header.

createCsvClassifierRequest_disableValueTrimming :: Lens' CreateCsvClassifierRequest (Maybe Bool)

Specifies not to trim values before identifying the type of column values. The default value is true.

createCsvClassifierRequest_header :: Lens' CreateCsvClassifierRequest (Maybe [Text])

A list of strings representing column names.

createCsvClassifierRequest_allowSingleColumn :: Lens' CreateCsvClassifierRequest (Maybe Bool)

Enables the processing of files that contain only one column.

createCsvClassifierRequest_delimiter :: Lens' CreateCsvClassifierRequest (Maybe Text)

A custom symbol to denote what separates each column entry in the row.

createCsvClassifierRequest_name :: Lens' CreateCsvClassifierRequest Text

The name of the classifier.