Network.AWS.Glue.Types.CsvClassifier

Description

Documentation

data CsvClassifier

A classifier for custom CSV content.

See: newCsvClassifier smart constructor.

Constructors

CsvClassifier'

Fields

creationTime :: Maybe POSIX

The time that this classifier was registered.

quoteSymbol :: Maybe Text

A custom symbol to denote what combines content into a single column value. It must be different from the column delimiter.

containsHeader :: Maybe CsvHeaderOption

Indicates whether the CSV file contains a header.

lastUpdated :: Maybe POSIX

The time that this classifier was last updated.

disableValueTrimming :: Maybe Bool

Specifies not to trim values before identifying the type of column values. The default value is true.

header :: Maybe [Text]

A list of strings representing column names.

version :: Maybe Integer

The version of this classifier.

allowSingleColumn :: Maybe Bool

Enables the processing of files that contain only one column.

delimiter :: Maybe Text

A custom symbol to denote what separates each column entry in the row.

name :: Text

The name of the classifier.

newCsvClassifier

Arguments

:: Text
$sel:name:CsvClassifier'

-> CsvClassifier

Create a value of CsvClassifier with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:CsvClassifier', csvClassifier_creationTime - The time that this classifier was registered.

$sel:quoteSymbol:CsvClassifier', csvClassifier_quoteSymbol - A custom symbol to denote what combines content into a single column value. It must be different from the column delimiter.

$sel:containsHeader:CsvClassifier', csvClassifier_containsHeader - Indicates whether the CSV file contains a header.

$sel:lastUpdated:CsvClassifier', csvClassifier_lastUpdated - The time that this classifier was last updated.

$sel:disableValueTrimming:CsvClassifier', csvClassifier_disableValueTrimming - Specifies not to trim values before identifying the type of column values. The default value is true.

$sel:header:CsvClassifier', csvClassifier_header - A list of strings representing column names.

$sel:version:CsvClassifier', csvClassifier_version - The version of this classifier.

$sel:allowSingleColumn:CsvClassifier', csvClassifier_allowSingleColumn - Enables the processing of files that contain only one column.

$sel:delimiter:CsvClassifier', csvClassifier_delimiter - A custom symbol to denote what separates each column entry in the row.

$sel:name:CsvClassifier', csvClassifier_name - The name of the classifier.

csvClassifier_creationTime :: Lens' CsvClassifier (Maybe UTCTime)

The time that this classifier was registered.

csvClassifier_quoteSymbol :: Lens' CsvClassifier (Maybe Text)

A custom symbol to denote what combines content into a single column value. It must be different from the column delimiter.

csvClassifier_containsHeader :: Lens' CsvClassifier (Maybe CsvHeaderOption)

Indicates whether the CSV file contains a header.

csvClassifier_lastUpdated :: Lens' CsvClassifier (Maybe UTCTime)

The time that this classifier was last updated.

csvClassifier_disableValueTrimming :: Lens' CsvClassifier (Maybe Bool)

Specifies not to trim values before identifying the type of column values. The default value is true.

csvClassifier_header :: Lens' CsvClassifier (Maybe [Text])

A list of strings representing column names.

csvClassifier_version :: Lens' CsvClassifier (Maybe Integer)

The version of this classifier.

csvClassifier_allowSingleColumn :: Lens' CsvClassifier (Maybe Bool)

Enables the processing of files that contain only one column.

csvClassifier_delimiter :: Lens' CsvClassifier (Maybe Text)

A custom symbol to denote what separates each column entry in the row.

csvClassifier_name :: Lens' CsvClassifier Text

The name of the classifier.