Network.AWS.Glue.Types.Database

Description

Documentation

data Database

The Database object represents a logical grouping of tables that might reside in a Hive metastore or an RDBMS.

See: newDatabase smart constructor.

Constructors

Database'

Fields

locationUri :: Maybe Text

The location of the database (for example, an HDFS path).

catalogId :: Maybe Text

The ID of the Data Catalog in which the database resides.

targetDatabase :: Maybe DatabaseIdentifier

A DatabaseIdentifier structure that describes a target database for resource linking.

parameters :: Maybe (HashMap Text Text)

These key-value pairs define parameters and properties of the database.

description :: Maybe Text

A description of the database.

createTime :: Maybe POSIX

The time at which the metadata database was created in the catalog.

createTableDefaultPermissions :: Maybe [PrincipalPermissions]

Creates a set of default permissions on the table for principals.

name :: Text

The name of the database. For Hive compatibility, this is folded to lowercase when it is stored.

newDatabase

Arguments

:: Text
$sel:name:Database'

-> Database

Create a value of Database with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:locationUri:Database', database_locationUri - The location of the database (for example, an HDFS path).

$sel:catalogId:Database', database_catalogId - The ID of the Data Catalog in which the database resides.

$sel:targetDatabase:Database', database_targetDatabase - A DatabaseIdentifier structure that describes a target database for resource linking.

$sel:parameters:Database', database_parameters - These key-value pairs define parameters and properties of the database.

$sel:description:Database', database_description - A description of the database.

$sel:createTime:Database', database_createTime - The time at which the metadata database was created in the catalog.

$sel:createTableDefaultPermissions:Database', database_createTableDefaultPermissions - Creates a set of default permissions on the table for principals.

$sel:name:Database', database_name - The name of the database. For Hive compatibility, this is folded to lowercase when it is stored.

database_locationUri :: Lens' Database (Maybe Text)

The location of the database (for example, an HDFS path).

database_catalogId :: Lens' Database (Maybe Text)

The ID of the Data Catalog in which the database resides.

database_targetDatabase :: Lens' Database (Maybe DatabaseIdentifier)

A DatabaseIdentifier structure that describes a target database for resource linking.

database_parameters :: Lens' Database (Maybe (HashMap Text Text))

These key-value pairs define parameters and properties of the database.

database_description :: Lens' Database (Maybe Text)

A description of the database.

database_createTime :: Lens' Database (Maybe UTCTime)

The time at which the metadata database was created in the catalog.

database_createTableDefaultPermissions :: Lens' Database (Maybe [PrincipalPermissions])

Creates a set of default permissions on the table for principals.

database_name :: Lens' Database Text

The name of the database. For Hive compatibility, this is folded to lowercase when it is stored.