Network.AWS.Glue.Types.DatabaseInput

Description

Documentation

data DatabaseInput

The structure used to create or update a database.

See: newDatabaseInput smart constructor.

Constructors

DatabaseInput'

Fields

createTableDefaultPermissions :: Maybe [PrincipalPermissions]

Creates a set of default permissions on the table for principals.

targetDatabase :: Maybe DatabaseIdentifier

A DatabaseIdentifier structure that describes a target database for resource linking.

description :: Maybe Text

A description of the database.

locationUri :: Maybe Text

The location of the database (for example, an HDFS path).

parameters :: Maybe (HashMap Text Text)

These key-value pairs define parameters and properties of the database.

These key-value pairs define parameters and properties of the database.

name :: Text

The name of the database. For Hive compatibility, this is folded to lowercase when it is stored.

newDatabaseInput

Arguments

:: Text
$sel:name:DatabaseInput'

-> DatabaseInput

Create a value of DatabaseInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createTableDefaultPermissions:DatabaseInput', databaseInput_createTableDefaultPermissions - Creates a set of default permissions on the table for principals.

$sel:targetDatabase:DatabaseInput', databaseInput_targetDatabase - A DatabaseIdentifier structure that describes a target database for resource linking.

$sel:description:DatabaseInput', databaseInput_description - A description of the database.

$sel:locationUri:DatabaseInput', databaseInput_locationUri - The location of the database (for example, an HDFS path).

$sel:parameters:DatabaseInput', databaseInput_parameters - These key-value pairs define parameters and properties of the database.

These key-value pairs define parameters and properties of the database.

$sel:name:DatabaseInput', databaseInput_name - The name of the database. For Hive compatibility, this is folded to lowercase when it is stored.

databaseInput_createTableDefaultPermissions :: Lens' DatabaseInput (Maybe [PrincipalPermissions])

Creates a set of default permissions on the table for principals.

databaseInput_targetDatabase :: Lens' DatabaseInput (Maybe DatabaseIdentifier)

A DatabaseIdentifier structure that describes a target database for resource linking.

databaseInput_description :: Lens' DatabaseInput (Maybe Text)

A description of the database.

databaseInput_locationUri :: Lens' DatabaseInput (Maybe Text)

The location of the database (for example, an HDFS path).

databaseInput_parameters :: Lens' DatabaseInput (Maybe (HashMap Text Text))

These key-value pairs define parameters and properties of the database.

These key-value pairs define parameters and properties of the database.

databaseInput_name :: Lens' DatabaseInput Text

The name of the database. For Hive compatibility, this is folded to lowercase when it is stored.