Network.AWS.Glue.Types.DevEndpoint

Description

Documentation

data DevEndpoint

A development endpoint where a developer can remotely debug extract, transform, and load (ETL) scripts.

See: newDevEndpoint smart constructor.

Constructors

DevEndpoint'

Fields

status :: Maybe Text

The current status of this DevEndpoint.

failureReason :: Maybe Text

The reason for a current failure in this DevEndpoint.

endpointName :: Maybe Text

The name of the DevEndpoint.

numberOfWorkers :: Maybe Int

The number of workers of a defined workerType that are allocated to the development endpoint.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

extraPythonLibsS3Path :: Maybe Text

The paths to one or more Python libraries in an Amazon S3 bucket that should be loaded in your DevEndpoint. Multiple values must be complete paths separated by a comma.

You can only use pure Python libraries with a DevEndpoint. Libraries that rely on C extensions, such as the pandas Python data analysis library, are not currently supported.

lastUpdateStatus :: Maybe Text

The status of the last update.

securityGroupIds :: Maybe [Text]

A list of security group identifiers used in this DevEndpoint.

lastModifiedTimestamp :: Maybe POSIX

The point in time at which this DevEndpoint was last modified.

publicKeys :: Maybe [Text]

A list of public keys to be used by the DevEndpoints for authentication. Using this attribute is preferred over a single public key because the public keys allow you to have a different private key per client.

If you previously created an endpoint with a public key, you must remove that key to be able to set a list of public keys. Call the UpdateDevEndpoint API operation with the public key content in the deletePublicKeys attribute, and the list of new keys in the addPublicKeys attribute.

vpcId :: Maybe Text

The ID of the virtual private cloud (VPC) used by this DevEndpoint.

arguments :: Maybe (HashMap Text Text)

A map of arguments used to configure the DevEndpoint.

Valid arguments are:

"--enable-glue-datacatalog": ""

You can specify a version of Python support for development endpoints by using the Arguments parameter in the CreateDevEndpoint or UpdateDevEndpoint APIs. If no arguments are provided, the version defaults to Python 2.

privateAddress :: Maybe Text

A private IP address to access the DevEndpoint within a VPC if the DevEndpoint is created within one. The PrivateAddress field is present only when you create the DevEndpoint within your VPC.

workerType :: Maybe WorkerType

The type of predefined worker that is allocated to the development endpoint. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker maps to 1 DPU (4 vCPU, 16 GB of memory, 64 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

For the G.2X worker type, each worker maps to 2 DPU (8 vCPU, 32 GB of memory, 128 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

Known issue: when a development endpoint is created with the G.2X WorkerType configuration, the Spark drivers for the development endpoint will run on 4 vCPU, 16 GB of memory, and a 64 GB disk.

securityConfiguration :: Maybe Text

The name of the SecurityConfiguration structure to be used with this DevEndpoint.

publicKey :: Maybe Text

The public key to be used by this DevEndpoint for authentication. This attribute is provided for backward compatibility because the recommended attribute to use is public keys.

subnetId :: Maybe Text

The subnet ID for this DevEndpoint.

glueVersion :: Maybe Text

Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for running your ETL scripts on development endpoints.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

Development endpoints that are created without specifying a Glue version default to Glue 0.9.

You can specify a version of Python support for development endpoints by using the Arguments parameter in the CreateDevEndpoint or UpdateDevEndpoint APIs. If no arguments are provided, the version defaults to Python 2.

numberOfNodes :: Maybe Int

The number of Glue Data Processing Units (DPUs) allocated to this DevEndpoint.

publicAddress :: Maybe Text

The public IP address used by this DevEndpoint. The PublicAddress field is present only when you create a non-virtual private cloud (VPC) DevEndpoint.

availabilityZone :: Maybe Text

The AWS Availability Zone where this DevEndpoint is located.

zeppelinRemoteSparkInterpreterPort :: Maybe Int

The Apache Zeppelin port for the remote Apache Spark interpreter.

extraJarsS3Path :: Maybe Text

The path to one or more Java .jar files in an S3 bucket that should be loaded in your DevEndpoint.

You can only use pure Java/Scala libraries with a DevEndpoint.

createdTimestamp :: Maybe POSIX

The point in time at which this DevEndpoint was created.

yarnEndpointAddress :: Maybe Text

The YARN endpoint address used by this DevEndpoint.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role used in this DevEndpoint.

newDevEndpoint :: DevEndpoint

Create a value of DevEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DevEndpoint', devEndpoint_status - The current status of this DevEndpoint.

$sel:failureReason:DevEndpoint', devEndpoint_failureReason - The reason for a current failure in this DevEndpoint.

$sel:endpointName:DevEndpoint', devEndpoint_endpointName - The name of the DevEndpoint.

$sel:numberOfWorkers:DevEndpoint', devEndpoint_numberOfWorkers - The number of workers of a defined workerType that are allocated to the development endpoint.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

$sel:extraPythonLibsS3Path:DevEndpoint', devEndpoint_extraPythonLibsS3Path - The paths to one or more Python libraries in an Amazon S3 bucket that should be loaded in your DevEndpoint. Multiple values must be complete paths separated by a comma.

You can only use pure Python libraries with a DevEndpoint. Libraries that rely on C extensions, such as the pandas Python data analysis library, are not currently supported.

$sel:lastUpdateStatus:DevEndpoint', devEndpoint_lastUpdateStatus - The status of the last update.

$sel:securityGroupIds:DevEndpoint', devEndpoint_securityGroupIds - A list of security group identifiers used in this DevEndpoint.

$sel:lastModifiedTimestamp:DevEndpoint', devEndpoint_lastModifiedTimestamp - The point in time at which this DevEndpoint was last modified.

$sel:publicKeys:DevEndpoint', devEndpoint_publicKeys - A list of public keys to be used by the DevEndpoints for authentication. Using this attribute is preferred over a single public key because the public keys allow you to have a different private key per client.

If you previously created an endpoint with a public key, you must remove that key to be able to set a list of public keys. Call the UpdateDevEndpoint API operation with the public key content in the deletePublicKeys attribute, and the list of new keys in the addPublicKeys attribute.

$sel:vpcId:DevEndpoint', devEndpoint_vpcId - The ID of the virtual private cloud (VPC) used by this DevEndpoint.

$sel:arguments:DevEndpoint', devEndpoint_arguments - A map of arguments used to configure the DevEndpoint.

Valid arguments are:

"--enable-glue-datacatalog": ""

You can specify a version of Python support for development endpoints by using the Arguments parameter in the CreateDevEndpoint or UpdateDevEndpoint APIs. If no arguments are provided, the version defaults to Python 2.

$sel:privateAddress:DevEndpoint', devEndpoint_privateAddress - A private IP address to access the DevEndpoint within a VPC if the DevEndpoint is created within one. The PrivateAddress field is present only when you create the DevEndpoint within your VPC.

$sel:workerType:DevEndpoint', devEndpoint_workerType - The type of predefined worker that is allocated to the development endpoint. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker maps to 1 DPU (4 vCPU, 16 GB of memory, 64 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

For the G.2X worker type, each worker maps to 2 DPU (8 vCPU, 32 GB of memory, 128 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

Known issue: when a development endpoint is created with the G.2X WorkerType configuration, the Spark drivers for the development endpoint will run on 4 vCPU, 16 GB of memory, and a 64 GB disk.

$sel:securityConfiguration:DevEndpoint', devEndpoint_securityConfiguration - The name of the SecurityConfiguration structure to be used with this DevEndpoint.

$sel:publicKey:DevEndpoint', devEndpoint_publicKey - The public key to be used by this DevEndpoint for authentication. This attribute is provided for backward compatibility because the recommended attribute to use is public keys.

$sel:subnetId:DevEndpoint', devEndpoint_subnetId - The subnet ID for this DevEndpoint.

$sel:glueVersion:DevEndpoint', devEndpoint_glueVersion - Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for running your ETL scripts on development endpoints.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

Development endpoints that are created without specifying a Glue version default to Glue 0.9.

You can specify a version of Python support for development endpoints by using the Arguments parameter in the CreateDevEndpoint or UpdateDevEndpoint APIs. If no arguments are provided, the version defaults to Python 2.

$sel:numberOfNodes:DevEndpoint', devEndpoint_numberOfNodes - The number of Glue Data Processing Units (DPUs) allocated to this DevEndpoint.

$sel:publicAddress:DevEndpoint', devEndpoint_publicAddress - The public IP address used by this DevEndpoint. The PublicAddress field is present only when you create a non-virtual private cloud (VPC) DevEndpoint.

$sel:availabilityZone:DevEndpoint', devEndpoint_availabilityZone - The AWS Availability Zone where this DevEndpoint is located.

$sel:zeppelinRemoteSparkInterpreterPort:DevEndpoint', devEndpoint_zeppelinRemoteSparkInterpreterPort - The Apache Zeppelin port for the remote Apache Spark interpreter.

$sel:extraJarsS3Path:DevEndpoint', devEndpoint_extraJarsS3Path - The path to one or more Java .jar files in an S3 bucket that should be loaded in your DevEndpoint.

You can only use pure Java/Scala libraries with a DevEndpoint.

$sel:createdTimestamp:DevEndpoint', devEndpoint_createdTimestamp - The point in time at which this DevEndpoint was created.

$sel:yarnEndpointAddress:DevEndpoint', devEndpoint_yarnEndpointAddress - The YARN endpoint address used by this DevEndpoint.

$sel:roleArn:DevEndpoint', devEndpoint_roleArn - The Amazon Resource Name (ARN) of the IAM role used in this DevEndpoint.

devEndpoint_status :: Lens' DevEndpoint (Maybe Text)

The current status of this DevEndpoint.

devEndpoint_failureReason :: Lens' DevEndpoint (Maybe Text)

The reason for a current failure in this DevEndpoint.

devEndpoint_endpointName :: Lens' DevEndpoint (Maybe Text)

The name of the DevEndpoint.

devEndpoint_numberOfWorkers :: Lens' DevEndpoint (Maybe Int)

The number of workers of a defined workerType that are allocated to the development endpoint.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

devEndpoint_extraPythonLibsS3Path :: Lens' DevEndpoint (Maybe Text)

The paths to one or more Python libraries in an Amazon S3 bucket that should be loaded in your DevEndpoint. Multiple values must be complete paths separated by a comma.

You can only use pure Python libraries with a DevEndpoint. Libraries that rely on C extensions, such as the pandas Python data analysis library, are not currently supported.

devEndpoint_lastUpdateStatus :: Lens' DevEndpoint (Maybe Text)

The status of the last update.

devEndpoint_securityGroupIds :: Lens' DevEndpoint (Maybe [Text])

A list of security group identifiers used in this DevEndpoint.

devEndpoint_lastModifiedTimestamp :: Lens' DevEndpoint (Maybe UTCTime)

The point in time at which this DevEndpoint was last modified.

devEndpoint_publicKeys :: Lens' DevEndpoint (Maybe [Text])

A list of public keys to be used by the DevEndpoints for authentication. Using this attribute is preferred over a single public key because the public keys allow you to have a different private key per client.

If you previously created an endpoint with a public key, you must remove that key to be able to set a list of public keys. Call the UpdateDevEndpoint API operation with the public key content in the deletePublicKeys attribute, and the list of new keys in the addPublicKeys attribute.

devEndpoint_vpcId :: Lens' DevEndpoint (Maybe Text)

The ID of the virtual private cloud (VPC) used by this DevEndpoint.

devEndpoint_arguments :: Lens' DevEndpoint (Maybe (HashMap Text Text))

A map of arguments used to configure the DevEndpoint.

Valid arguments are:

"--enable-glue-datacatalog": ""

You can specify a version of Python support for development endpoints by using the Arguments parameter in the CreateDevEndpoint or UpdateDevEndpoint APIs. If no arguments are provided, the version defaults to Python 2.

devEndpoint_privateAddress :: Lens' DevEndpoint (Maybe Text)

A private IP address to access the DevEndpoint within a VPC if the DevEndpoint is created within one. The PrivateAddress field is present only when you create the DevEndpoint within your VPC.

devEndpoint_workerType :: Lens' DevEndpoint (Maybe WorkerType)

The type of predefined worker that is allocated to the development endpoint. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker maps to 1 DPU (4 vCPU, 16 GB of memory, 64 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

For the G.2X worker type, each worker maps to 2 DPU (8 vCPU, 32 GB of memory, 128 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

Known issue: when a development endpoint is created with the G.2X WorkerType configuration, the Spark drivers for the development endpoint will run on 4 vCPU, 16 GB of memory, and a 64 GB disk.

devEndpoint_securityConfiguration :: Lens' DevEndpoint (Maybe Text)

The name of the SecurityConfiguration structure to be used with this DevEndpoint.

devEndpoint_publicKey :: Lens' DevEndpoint (Maybe Text)

The public key to be used by this DevEndpoint for authentication. This attribute is provided for backward compatibility because the recommended attribute to use is public keys.

devEndpoint_subnetId :: Lens' DevEndpoint (Maybe Text)

The subnet ID for this DevEndpoint.

devEndpoint_glueVersion :: Lens' DevEndpoint (Maybe Text)

Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for running your ETL scripts on development endpoints.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

Development endpoints that are created without specifying a Glue version default to Glue 0.9.

You can specify a version of Python support for development endpoints by using the Arguments parameter in the CreateDevEndpoint or UpdateDevEndpoint APIs. If no arguments are provided, the version defaults to Python 2.

devEndpoint_numberOfNodes :: Lens' DevEndpoint (Maybe Int)

The number of Glue Data Processing Units (DPUs) allocated to this DevEndpoint.

devEndpoint_publicAddress :: Lens' DevEndpoint (Maybe Text)

The public IP address used by this DevEndpoint. The PublicAddress field is present only when you create a non-virtual private cloud (VPC) DevEndpoint.

devEndpoint_availabilityZone :: Lens' DevEndpoint (Maybe Text)

The AWS Availability Zone where this DevEndpoint is located.

devEndpoint_zeppelinRemoteSparkInterpreterPort :: Lens' DevEndpoint (Maybe Int)

The Apache Zeppelin port for the remote Apache Spark interpreter.

devEndpoint_extraJarsS3Path :: Lens' DevEndpoint (Maybe Text)

The path to one or more Java .jar files in an S3 bucket that should be loaded in your DevEndpoint.

You can only use pure Java/Scala libraries with a DevEndpoint.

devEndpoint_createdTimestamp :: Lens' DevEndpoint (Maybe UTCTime)

The point in time at which this DevEndpoint was created.

devEndpoint_yarnEndpointAddress :: Lens' DevEndpoint (Maybe Text)

The YARN endpoint address used by this DevEndpoint.

devEndpoint_roleArn :: Lens' DevEndpoint (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role used in this DevEndpoint.