Network.AWS.Glue.Types.FindMatchesMetrics

Description

Documentation

data FindMatchesMetrics

The evaluation metrics for the find matches algorithm. The quality of your machine learning transform is measured by getting your transform to predict some matches and comparing the results to known matches from the same dataset. The quality metrics are based on a subset of your data, so they are not precise.

See: newFindMatchesMetrics smart constructor.

Constructors

FindMatchesMetrics'

Fields

f1 :: Maybe Double

The maximum F1 metric indicates the transform's accuracy between 0 and 1, where 1 is the best accuracy.

For more information, see F1 score in Wikipedia.

columnImportances :: Maybe [ColumnImportance]

A list of ColumnImportance structures containing column importance metrics, sorted in order of descending importance.

confusionMatrix :: Maybe ConfusionMatrix

The confusion matrix shows you what your transform is predicting accurately and what types of errors it is making.

For more information, see Confusion matrix in Wikipedia.

precision :: Maybe Double

The precision metric indicates when often your transform is correct when it predicts a match. Specifically, it measures how well the transform finds true positives from the total true positives possible.

For more information, see Precision and recall in Wikipedia.

areaUnderPRCurve :: Maybe Double

The area under the precision/recall curve (AUPRC) is a single number measuring the overall quality of the transform, that is independent of the choice made for precision vs. recall. Higher values indicate that you have a more attractive precision vs. recall tradeoff.

For more information, see Precision and recall in Wikipedia.

recall :: Maybe Double

The recall metric indicates that for an actual match, how often your transform predicts the match. Specifically, it measures how well the transform finds true positives from the total records in the source data.

For more information, see Precision and recall in Wikipedia.

newFindMatchesMetrics :: FindMatchesMetrics

Create a value of FindMatchesMetrics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:f1:FindMatchesMetrics', findMatchesMetrics_f1 - The maximum F1 metric indicates the transform's accuracy between 0 and 1, where 1 is the best accuracy.

For more information, see F1 score in Wikipedia.

$sel:columnImportances:FindMatchesMetrics', findMatchesMetrics_columnImportances - A list of ColumnImportance structures containing column importance metrics, sorted in order of descending importance.

$sel:confusionMatrix:FindMatchesMetrics', findMatchesMetrics_confusionMatrix - The confusion matrix shows you what your transform is predicting accurately and what types of errors it is making.

For more information, see Confusion matrix in Wikipedia.

$sel:precision:FindMatchesMetrics', findMatchesMetrics_precision - The precision metric indicates when often your transform is correct when it predicts a match. Specifically, it measures how well the transform finds true positives from the total true positives possible.

For more information, see Precision and recall in Wikipedia.

$sel:areaUnderPRCurve:FindMatchesMetrics', findMatchesMetrics_areaUnderPRCurve - The area under the precision/recall curve (AUPRC) is a single number measuring the overall quality of the transform, that is independent of the choice made for precision vs. recall. Higher values indicate that you have a more attractive precision vs. recall tradeoff.

For more information, see Precision and recall in Wikipedia.

$sel:recall:FindMatchesMetrics', findMatchesMetrics_recall - The recall metric indicates that for an actual match, how often your transform predicts the match. Specifically, it measures how well the transform finds true positives from the total records in the source data.

For more information, see Precision and recall in Wikipedia.

findMatchesMetrics_f1 :: Lens' FindMatchesMetrics (Maybe Double)

The maximum F1 metric indicates the transform's accuracy between 0 and 1, where 1 is the best accuracy.

For more information, see F1 score in Wikipedia.

findMatchesMetrics_columnImportances :: Lens' FindMatchesMetrics (Maybe [ColumnImportance])

A list of ColumnImportance structures containing column importance metrics, sorted in order of descending importance.

findMatchesMetrics_confusionMatrix :: Lens' FindMatchesMetrics (Maybe ConfusionMatrix)

The confusion matrix shows you what your transform is predicting accurately and what types of errors it is making.

For more information, see Confusion matrix in Wikipedia.

findMatchesMetrics_precision :: Lens' FindMatchesMetrics (Maybe Double)

The precision metric indicates when often your transform is correct when it predicts a match. Specifically, it measures how well the transform finds true positives from the total true positives possible.

For more information, see Precision and recall in Wikipedia.

findMatchesMetrics_areaUnderPRCurve :: Lens' FindMatchesMetrics (Maybe Double)

The area under the precision/recall curve (AUPRC) is a single number measuring the overall quality of the transform, that is independent of the choice made for precision vs. recall. Higher values indicate that you have a more attractive precision vs. recall tradeoff.

For more information, see Precision and recall in Wikipedia.

findMatchesMetrics_recall :: Lens' FindMatchesMetrics (Maybe Double)

The recall metric indicates that for an actual match, how often your transform predicts the match. Specifically, it measures how well the transform finds true positives from the total records in the source data.

For more information, see Precision and recall in Wikipedia.