Network.AWS.Glue.Types.GlueTable

Description

Documentation

data GlueTable

The database and table in the Glue Data Catalog that is used for input or output data.

See: newGlueTable smart constructor.

Constructors

GlueTable'

Fields

connectionName :: Maybe Text

The name of the connection to the Glue Data Catalog.

catalogId :: Maybe Text

A unique identifier for the Glue Data Catalog.

databaseName :: Text

A database name in the Glue Data Catalog.

tableName :: Text

A table name in the Glue Data Catalog.

newGlueTable

Arguments

:: Text
$sel:databaseName:GlueTable'

-> Text
$sel:tableName:GlueTable'

-> GlueTable

Create a value of GlueTable with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectionName:GlueTable', glueTable_connectionName - The name of the connection to the Glue Data Catalog.

$sel:catalogId:GlueTable', glueTable_catalogId - A unique identifier for the Glue Data Catalog.

$sel:databaseName:GlueTable', glueTable_databaseName - A database name in the Glue Data Catalog.

$sel:tableName:GlueTable', glueTable_tableName - A table name in the Glue Data Catalog.

glueTable_connectionName :: Lens' GlueTable (Maybe Text)

The name of the connection to the Glue Data Catalog.

glueTable_catalogId :: Lens' GlueTable (Maybe Text)

A unique identifier for the Glue Data Catalog.

glueTable_databaseName :: Lens' GlueTable Text

A database name in the Glue Data Catalog.

glueTable_tableName :: Lens' GlueTable Text

A table name in the Glue Data Catalog.