Network.AWS.Glue.Types.Job

Description

Documentation

data Job

Specifies a job definition.

See: newJob smart constructor.

Constructors

Job'

Fields

nonOverridableArguments :: Maybe (HashMap Text Text)

Non-overridable arguments for this job, specified as name-value pairs.

createdOn :: Maybe POSIX

The time and date that this job definition was created.

securityConfiguration :: Maybe Text

The name of the SecurityConfiguration structure to be used with this job.

timeout :: Maybe Natural

The job timeout in minutes. This is the maximum time that a job run can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours).

maxCapacity :: Maybe Double

For Glue version 1.0 or earlier jobs, using the standard worker type, the number of Glue data processing units (DPUs) that can be allocated when this job runs. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

Do not set Max Capacity if using WorkerType and NumberOfWorkers.

The value that can be allocated for MaxCapacity depends on whether you are running a Python shell job, an Apache Spark ETL job, or an Apache Spark streaming ETL job:

When you specify a Python shell job (JobCommand.Name="pythonshell"), you can allocate either 0.0625 or 1 DPU. The default is 0.0625 DPU.

When you specify an Apache Spark ETL job (JobCommand.Name="glueetl") or Apache Spark streaming ETL job (JobCommand.Name="gluestreaming"), you can allocate from 2 to 100 DPUs. The default is 10 DPUs. This job type cannot have a fractional DPU allocation.

For Glue version 2.0 jobs, you cannot instead specify a Maximum capacity. Instead, you should specify a Worker type and the Number of workers.

connections :: Maybe ConnectionsList

The connections used for this job.

lastModifiedOn :: Maybe POSIX

The last point in time when this job definition was modified.

notificationProperty :: Maybe NotificationProperty

Specifies configuration properties of a job notification.

numberOfWorkers :: Maybe Int

The number of workers of a defined workerType that are allocated when a job runs.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

name :: Maybe Text

The name you assign to this job definition.

command :: Maybe JobCommand

The JobCommand that runs this job.

role' :: Maybe Text

The name or Amazon Resource Name (ARN) of the IAM role associated with this job.

glueVersion :: Maybe Text

Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for jobs of type Spark.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

Jobs that are created without specifying a Glue version default to Glue 0.9.

workerType :: Maybe WorkerType

The type of predefined worker that is allocated when a job runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker maps to 1 DPU (4 vCPU, 16 GB of memory, 64 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

For the G.2X worker type, each worker maps to 2 DPU (8 vCPU, 32 GB of memory, 128 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

description :: Maybe Text

A description of the job.

defaultArguments :: Maybe (HashMap Text Text)

The default arguments for this job, specified as name-value pairs.

You can specify arguments here that your own job-execution script consumes, as well as arguments that Glue itself consumes.

For information about how to specify and consume your own Job arguments, see the Calling Glue APIs in Python topic in the developer guide.

For information about the key-value pairs that Glue consumes to set up your job, see the Special Parameters Used by Glue topic in the developer guide.

executionProperty :: Maybe ExecutionProperty

An ExecutionProperty specifying the maximum number of concurrent runs allowed for this job.

allocatedCapacity :: Maybe Int

This field is deprecated. Use MaxCapacity instead.

The number of Glue data processing units (DPUs) allocated to runs of this job. You can allocate from 2 to 100 DPUs; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

logUri :: Maybe Text

This field is reserved for future use.

maxRetries :: Maybe Int

The maximum number of times to retry this job after a JobRun fails.

newJob :: Job

Create a value of Job with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nonOverridableArguments:Job', job_nonOverridableArguments - Non-overridable arguments for this job, specified as name-value pairs.

$sel:createdOn:Job', job_createdOn - The time and date that this job definition was created.

$sel:securityConfiguration:Job', job_securityConfiguration - The name of the SecurityConfiguration structure to be used with this job.

$sel:timeout:Job', job_timeout - The job timeout in minutes. This is the maximum time that a job run can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours).

$sel:maxCapacity:Job', job_maxCapacity - For Glue version 1.0 or earlier jobs, using the standard worker type, the number of Glue data processing units (DPUs) that can be allocated when this job runs. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

Do not set Max Capacity if using WorkerType and NumberOfWorkers.

The value that can be allocated for MaxCapacity depends on whether you are running a Python shell job, an Apache Spark ETL job, or an Apache Spark streaming ETL job:

When you specify a Python shell job (JobCommand.Name="pythonshell"), you can allocate either 0.0625 or 1 DPU. The default is 0.0625 DPU.

When you specify an Apache Spark ETL job (JobCommand.Name="glueetl") or Apache Spark streaming ETL job (JobCommand.Name="gluestreaming"), you can allocate from 2 to 100 DPUs. The default is 10 DPUs. This job type cannot have a fractional DPU allocation.

For Glue version 2.0 jobs, you cannot instead specify a Maximum capacity. Instead, you should specify a Worker type and the Number of workers.

$sel:connections:Job', job_connections - The connections used for this job.

$sel:lastModifiedOn:Job', job_lastModifiedOn - The last point in time when this job definition was modified.

$sel:notificationProperty:Job', job_notificationProperty - Specifies configuration properties of a job notification.

$sel:numberOfWorkers:Job', job_numberOfWorkers - The number of workers of a defined workerType that are allocated when a job runs.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

$sel:name:Job', job_name - The name you assign to this job definition.

$sel:command:Job', job_command - The JobCommand that runs this job.

$sel:role':Job', job_role - The name or Amazon Resource Name (ARN) of the IAM role associated with this job.

$sel:glueVersion:Job', job_glueVersion - Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for jobs of type Spark.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

Jobs that are created without specifying a Glue version default to Glue 0.9.

$sel:workerType:Job', job_workerType - The type of predefined worker that is allocated when a job runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker maps to 1 DPU (4 vCPU, 16 GB of memory, 64 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

For the G.2X worker type, each worker maps to 2 DPU (8 vCPU, 32 GB of memory, 128 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

$sel:description:Job', job_description - A description of the job.

$sel:defaultArguments:Job', job_defaultArguments - The default arguments for this job, specified as name-value pairs.

You can specify arguments here that your own job-execution script consumes, as well as arguments that Glue itself consumes.

For information about how to specify and consume your own Job arguments, see the Calling Glue APIs in Python topic in the developer guide.

For information about the key-value pairs that Glue consumes to set up your job, see the Special Parameters Used by Glue topic in the developer guide.

$sel:executionProperty:Job', job_executionProperty - An ExecutionProperty specifying the maximum number of concurrent runs allowed for this job.

$sel:allocatedCapacity:Job', job_allocatedCapacity - This field is deprecated. Use MaxCapacity instead.

The number of Glue data processing units (DPUs) allocated to runs of this job. You can allocate from 2 to 100 DPUs; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

$sel:logUri:Job', job_logUri - This field is reserved for future use.

$sel:maxRetries:Job', job_maxRetries - The maximum number of times to retry this job after a JobRun fails.

job_nonOverridableArguments :: Lens' Job (Maybe (HashMap Text Text))

Non-overridable arguments for this job, specified as name-value pairs.

job_createdOn :: Lens' Job (Maybe UTCTime)

The time and date that this job definition was created.

job_securityConfiguration :: Lens' Job (Maybe Text)

The name of the SecurityConfiguration structure to be used with this job.

job_timeout :: Lens' Job (Maybe Natural)

The job timeout in minutes. This is the maximum time that a job run can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours).

job_maxCapacity :: Lens' Job (Maybe Double)

For Glue version 1.0 or earlier jobs, using the standard worker type, the number of Glue data processing units (DPUs) that can be allocated when this job runs. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

Do not set Max Capacity if using WorkerType and NumberOfWorkers.

The value that can be allocated for MaxCapacity depends on whether you are running a Python shell job, an Apache Spark ETL job, or an Apache Spark streaming ETL job:

When you specify a Python shell job (JobCommand.Name="pythonshell"), you can allocate either 0.0625 or 1 DPU. The default is 0.0625 DPU.

When you specify an Apache Spark ETL job (JobCommand.Name="glueetl") or Apache Spark streaming ETL job (JobCommand.Name="gluestreaming"), you can allocate from 2 to 100 DPUs. The default is 10 DPUs. This job type cannot have a fractional DPU allocation.

For Glue version 2.0 jobs, you cannot instead specify a Maximum capacity. Instead, you should specify a Worker type and the Number of workers.

job_connections :: Lens' Job (Maybe ConnectionsList)

The connections used for this job.

job_lastModifiedOn :: Lens' Job (Maybe UTCTime)

The last point in time when this job definition was modified.

job_notificationProperty :: Lens' Job (Maybe NotificationProperty)

Specifies configuration properties of a job notification.

job_numberOfWorkers :: Lens' Job (Maybe Int)

The number of workers of a defined workerType that are allocated when a job runs.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

job_name :: Lens' Job (Maybe Text)

The name you assign to this job definition.

job_command :: Lens' Job (Maybe JobCommand)

The JobCommand that runs this job.

job_role :: Lens' Job (Maybe Text)

The name or Amazon Resource Name (ARN) of the IAM role associated with this job.

job_glueVersion :: Lens' Job (Maybe Text)

Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for jobs of type Spark.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

Jobs that are created without specifying a Glue version default to Glue 0.9.

job_workerType :: Lens' Job (Maybe WorkerType)

The type of predefined worker that is allocated when a job runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker maps to 1 DPU (4 vCPU, 16 GB of memory, 64 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

For the G.2X worker type, each worker maps to 2 DPU (8 vCPU, 32 GB of memory, 128 GB disk), and provides 1 executor per worker. We recommend this worker type for memory-intensive jobs.

job_description :: Lens' Job (Maybe Text)

A description of the job.

job_defaultArguments :: Lens' Job (Maybe (HashMap Text Text))

The default arguments for this job, specified as name-value pairs.

You can specify arguments here that your own job-execution script consumes, as well as arguments that Glue itself consumes.

For information about how to specify and consume your own Job arguments, see the Calling Glue APIs in Python topic in the developer guide.

For information about the key-value pairs that Glue consumes to set up your job, see the Special Parameters Used by Glue topic in the developer guide.

job_executionProperty :: Lens' Job (Maybe ExecutionProperty)

An ExecutionProperty specifying the maximum number of concurrent runs allowed for this job.

job_allocatedCapacity :: Lens' Job (Maybe Int)

This field is deprecated. Use MaxCapacity instead.

The number of Glue data processing units (DPUs) allocated to runs of this job. You can allocate from 2 to 100 DPUs; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

job_logUri :: Lens' Job (Maybe Text)

This field is reserved for future use.

job_maxRetries :: Lens' Job (Maybe Int)

The maximum number of times to retry this job after a JobRun fails.