Network.AWS.Glue.Types.JobBookmarkEntry

Description

Documentation

data JobBookmarkEntry

Defines a point that a job can resume processing.

See: newJobBookmarkEntry smart constructor.

Constructors

JobBookmarkEntry'

Fields

runId :: Maybe Text

The run ID number.

jobBookmark :: Maybe Text

The bookmark itself.

version :: Maybe Int

The version of the job.

run :: Maybe Int

The run ID number.

jobName :: Maybe Text

The name of the job in question.

attempt :: Maybe Int

The attempt ID number.

previousRunId :: Maybe Text

The unique run identifier associated with the previous job run.

newJobBookmarkEntry :: JobBookmarkEntry

Create a value of JobBookmarkEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:runId:JobBookmarkEntry', jobBookmarkEntry_runId - The run ID number.

$sel:jobBookmark:JobBookmarkEntry', jobBookmarkEntry_jobBookmark - The bookmark itself.

$sel:version:JobBookmarkEntry', jobBookmarkEntry_version - The version of the job.

$sel:run:JobBookmarkEntry', jobBookmarkEntry_run - The run ID number.

$sel:jobName:JobBookmarkEntry', jobBookmarkEntry_jobName - The name of the job in question.

$sel:attempt:JobBookmarkEntry', jobBookmarkEntry_attempt - The attempt ID number.

$sel:previousRunId:JobBookmarkEntry', jobBookmarkEntry_previousRunId - The unique run identifier associated with the previous job run.

jobBookmarkEntry_runId :: Lens' JobBookmarkEntry (Maybe Text)

The run ID number.

jobBookmarkEntry_jobBookmark :: Lens' JobBookmarkEntry (Maybe Text)

The bookmark itself.

jobBookmarkEntry_version :: Lens' JobBookmarkEntry (Maybe Int)

The version of the job.

jobBookmarkEntry_run :: Lens' JobBookmarkEntry (Maybe Int)

The run ID number.

jobBookmarkEntry_jobName :: Lens' JobBookmarkEntry (Maybe Text)

The name of the job in question.

jobBookmarkEntry_attempt :: Lens' JobBookmarkEntry (Maybe Int)

The attempt ID number.

jobBookmarkEntry_previousRunId :: Lens' JobBookmarkEntry (Maybe Text)

The unique run identifier associated with the previous job run.