Network.AWS.Glue.Types.JobRun

Description

Documentation

data JobRun

Contains information about a job run.

See: newJobRun smart constructor.

Constructors

JobRun'

Fields

completedOn :: Maybe POSIX

The date and time that this job run completed.

numberOfWorkers :: Maybe Int

The number of workers of a defined workerType that are allocated when a job runs.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

triggerName :: Maybe Text

The name of the trigger that started this job run.

notificationProperty :: Maybe NotificationProperty

Specifies configuration properties of a job run notification.

lastModifiedOn :: Maybe POSIX

The last time that this job run was modified.

arguments :: Maybe (HashMap Text Text)

The job arguments associated with this run. For this job run, they replace the default arguments set in the job definition itself.

You can specify arguments here that your own job-execution script consumes, as well as arguments that Glue itself consumes.

For information about how to specify and consume your own job arguments, see the Calling Glue APIs in Python topic in the developer guide.

For information about the key-value pairs that Glue consumes to set up your job, see the Special Parameters Used by Glue topic in the developer guide.

jobName :: Maybe Text

The name of the job definition being used in this run.

startedOn :: Maybe POSIX

The date and time at which this job run was started.

workerType :: Maybe WorkerType

The type of predefined worker that is allocated when a job runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker provides 4 vCPU, 16 GB of memory and a 64GB disk, and 1 executor per worker.

For the G.2X worker type, each worker provides 8 vCPU, 32 GB of memory and a 128GB disk, and 1 executor per worker.

securityConfiguration :: Maybe Text

The name of the SecurityConfiguration structure to be used with this job run.

glueVersion :: Maybe Text

Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for jobs of type Spark.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

Jobs that are created without specifying a Glue version default to Glue 0.9.

jobRunState :: Maybe JobRunState

The current state of the job run. For more information about the statuses of jobs that have terminated abnormally, see Glue Job Run Statuses.

logGroupName :: Maybe Text

The name of the log group for secure logging that can be server-side encrypted in Amazon CloudWatch using KMS. This name can be /aws-glue/jobs/, in which case the default encryption is NONE. If you add a role name and SecurityConfiguration name (in other words, /aws-glue/jobs-yourRoleName-yourSecurityConfigurationName/), then that security configuration is used to encrypt the log group.

executionTime :: Maybe Int

The amount of time (in seconds) that the job run consumed resources.

predecessorRuns :: Maybe [Predecessor]

A list of predecessors to this job run.

previousRunId :: Maybe Text

The ID of the previous run of this job. For example, the JobRunId specified in the StartJobRun action.

id :: Maybe Text

The ID of this job run.

attempt :: Maybe Int

The number of the attempt to run this job.

allocatedCapacity :: Maybe Int

This field is deprecated. Use MaxCapacity instead.

The number of Glue data processing units (DPUs) allocated to this JobRun. From 2 to 100 DPUs can be allocated; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

maxCapacity :: Maybe Double

The number of Glue data processing units (DPUs) that can be allocated when this job runs. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

Do not set Max Capacity if using WorkerType and NumberOfWorkers.

The value that can be allocated for MaxCapacity depends on whether you are running a Python shell job or an Apache Spark ETL job:

When you specify a Python shell job (JobCommand.Name="pythonshell"), you can allocate either 0.0625 or 1 DPU. The default is 0.0625 DPU.

When you specify an Apache Spark ETL job (JobCommand.Name="glueetl"), you can allocate from 2 to 100 DPUs. The default is 10 DPUs. This job type cannot have a fractional DPU allocation.

timeout :: Maybe Natural

The JobRun timeout in minutes. This is the maximum time that a job run can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours). This overrides the timeout value set in the parent job.

errorMessage :: Maybe Text

An error message associated with this job run.

newJobRun :: JobRun

Create a value of JobRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:completedOn:JobRun', jobRun_completedOn - The date and time that this job run completed.

$sel:numberOfWorkers:JobRun', jobRun_numberOfWorkers - The number of workers of a defined workerType that are allocated when a job runs.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

$sel:triggerName:JobRun', jobRun_triggerName - The name of the trigger that started this job run.

$sel:notificationProperty:JobRun', jobRun_notificationProperty - Specifies configuration properties of a job run notification.

$sel:lastModifiedOn:JobRun', jobRun_lastModifiedOn - The last time that this job run was modified.

$sel:arguments:JobRun', jobRun_arguments - The job arguments associated with this run. For this job run, they replace the default arguments set in the job definition itself.

You can specify arguments here that your own job-execution script consumes, as well as arguments that Glue itself consumes.

For information about how to specify and consume your own job arguments, see the Calling Glue APIs in Python topic in the developer guide.

For information about the key-value pairs that Glue consumes to set up your job, see the Special Parameters Used by Glue topic in the developer guide.

$sel:jobName:JobRun', jobRun_jobName - The name of the job definition being used in this run.

$sel:startedOn:JobRun', jobRun_startedOn - The date and time at which this job run was started.

$sel:workerType:JobRun', jobRun_workerType - The type of predefined worker that is allocated when a job runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker provides 4 vCPU, 16 GB of memory and a 64GB disk, and 1 executor per worker.

For the G.2X worker type, each worker provides 8 vCPU, 32 GB of memory and a 128GB disk, and 1 executor per worker.

$sel:securityConfiguration:JobRun', jobRun_securityConfiguration - The name of the SecurityConfiguration structure to be used with this job run.

$sel:glueVersion:JobRun', jobRun_glueVersion - Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for jobs of type Spark.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

Jobs that are created without specifying a Glue version default to Glue 0.9.

$sel:jobRunState:JobRun', jobRun_jobRunState - The current state of the job run. For more information about the statuses of jobs that have terminated abnormally, see Glue Job Run Statuses.

$sel:logGroupName:JobRun', jobRun_logGroupName - The name of the log group for secure logging that can be server-side encrypted in Amazon CloudWatch using KMS. This name can be /aws-glue/jobs/, in which case the default encryption is NONE. If you add a role name and SecurityConfiguration name (in other words, /aws-glue/jobs-yourRoleName-yourSecurityConfigurationName/), then that security configuration is used to encrypt the log group.

$sel:executionTime:JobRun', jobRun_executionTime - The amount of time (in seconds) that the job run consumed resources.

$sel:predecessorRuns:JobRun', jobRun_predecessorRuns - A list of predecessors to this job run.

$sel:previousRunId:JobRun', jobRun_previousRunId - The ID of the previous run of this job. For example, the JobRunId specified in the StartJobRun action.

$sel:id:JobRun', jobRun_id - The ID of this job run.

$sel:attempt:JobRun', jobRun_attempt - The number of the attempt to run this job.

$sel:allocatedCapacity:JobRun', jobRun_allocatedCapacity - This field is deprecated. Use MaxCapacity instead.

The number of Glue data processing units (DPUs) allocated to this JobRun. From 2 to 100 DPUs can be allocated; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

$sel:maxCapacity:JobRun', jobRun_maxCapacity - The number of Glue data processing units (DPUs) that can be allocated when this job runs. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

Do not set Max Capacity if using WorkerType and NumberOfWorkers.

The value that can be allocated for MaxCapacity depends on whether you are running a Python shell job or an Apache Spark ETL job:

When you specify a Python shell job (JobCommand.Name="pythonshell"), you can allocate either 0.0625 or 1 DPU. The default is 0.0625 DPU.

When you specify an Apache Spark ETL job (JobCommand.Name="glueetl"), you can allocate from 2 to 100 DPUs. The default is 10 DPUs. This job type cannot have a fractional DPU allocation.

$sel:timeout:JobRun', jobRun_timeout - The JobRun timeout in minutes. This is the maximum time that a job run can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours). This overrides the timeout value set in the parent job.

$sel:errorMessage:JobRun', jobRun_errorMessage - An error message associated with this job run.

jobRun_completedOn :: Lens' JobRun (Maybe UTCTime)

The date and time that this job run completed.

jobRun_numberOfWorkers :: Lens' JobRun (Maybe Int)

The number of workers of a defined workerType that are allocated when a job runs.

The maximum number of workers you can define are 299 for G.1X, and 149 for G.2X.

jobRun_triggerName :: Lens' JobRun (Maybe Text)

The name of the trigger that started this job run.

jobRun_notificationProperty :: Lens' JobRun (Maybe NotificationProperty)

Specifies configuration properties of a job run notification.

jobRun_lastModifiedOn :: Lens' JobRun (Maybe UTCTime)

The last time that this job run was modified.

jobRun_arguments :: Lens' JobRun (Maybe (HashMap Text Text))

The job arguments associated with this run. For this job run, they replace the default arguments set in the job definition itself.

You can specify arguments here that your own job-execution script consumes, as well as arguments that Glue itself consumes.

For information about how to specify and consume your own job arguments, see the Calling Glue APIs in Python topic in the developer guide.

For information about the key-value pairs that Glue consumes to set up your job, see the Special Parameters Used by Glue topic in the developer guide.

jobRun_jobName :: Lens' JobRun (Maybe Text)

The name of the job definition being used in this run.

jobRun_startedOn :: Lens' JobRun (Maybe UTCTime)

The date and time at which this job run was started.

jobRun_workerType :: Lens' JobRun (Maybe WorkerType)

The type of predefined worker that is allocated when a job runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker provides 4 vCPU, 16 GB of memory and a 64GB disk, and 1 executor per worker.

For the G.2X worker type, each worker provides 8 vCPU, 32 GB of memory and a 128GB disk, and 1 executor per worker.

jobRun_securityConfiguration :: Lens' JobRun (Maybe Text)

The name of the SecurityConfiguration structure to be used with this job run.

jobRun_glueVersion :: Lens' JobRun (Maybe Text)

Glue version determines the versions of Apache Spark and Python that Glue supports. The Python version indicates the version supported for jobs of type Spark.

For more information about the available Glue versions and corresponding Spark and Python versions, see Glue version in the developer guide.

Jobs that are created without specifying a Glue version default to Glue 0.9.

jobRun_jobRunState :: Lens' JobRun (Maybe JobRunState)

The current state of the job run. For more information about the statuses of jobs that have terminated abnormally, see Glue Job Run Statuses.

jobRun_logGroupName :: Lens' JobRun (Maybe Text)

The name of the log group for secure logging that can be server-side encrypted in Amazon CloudWatch using KMS. This name can be /aws-glue/jobs/, in which case the default encryption is NONE. If you add a role name and SecurityConfiguration name (in other words, /aws-glue/jobs-yourRoleName-yourSecurityConfigurationName/), then that security configuration is used to encrypt the log group.

jobRun_executionTime :: Lens' JobRun (Maybe Int)

The amount of time (in seconds) that the job run consumed resources.

jobRun_predecessorRuns :: Lens' JobRun (Maybe [Predecessor])

A list of predecessors to this job run.

jobRun_previousRunId :: Lens' JobRun (Maybe Text)

The ID of the previous run of this job. For example, the JobRunId specified in the StartJobRun action.

jobRun_id :: Lens' JobRun (Maybe Text)

The ID of this job run.

jobRun_attempt :: Lens' JobRun (Maybe Int)

The number of the attempt to run this job.

jobRun_allocatedCapacity :: Lens' JobRun (Maybe Int)

This field is deprecated. Use MaxCapacity instead.

The number of Glue data processing units (DPUs) allocated to this JobRun. From 2 to 100 DPUs can be allocated; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

jobRun_maxCapacity :: Lens' JobRun (Maybe Double)

The number of Glue data processing units (DPUs) that can be allocated when this job runs. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

Do not set Max Capacity if using WorkerType and NumberOfWorkers.

The value that can be allocated for MaxCapacity depends on whether you are running a Python shell job or an Apache Spark ETL job:

When you specify a Python shell job (JobCommand.Name="pythonshell"), you can allocate either 0.0625 or 1 DPU. The default is 0.0625 DPU.

When you specify an Apache Spark ETL job (JobCommand.Name="glueetl"), you can allocate from 2 to 100 DPUs. The default is 10 DPUs. This job type cannot have a fractional DPU allocation.

jobRun_timeout :: Lens' JobRun (Maybe Natural)

The JobRun timeout in minutes. This is the maximum time that a job run can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours). This overrides the timeout value set in the parent job.

jobRun_errorMessage :: Lens' JobRun (Maybe Text)

An error message associated with this job run.