Network.AWS.Glue.Types.LastActiveDefinition

Description

Documentation

data LastActiveDefinition

When there are multiple versions of a blueprint and the latest version has some errors, this attribute indicates the last successful blueprint definition that is available with the service.

See: newLastActiveDefinition smart constructor.

Constructors

LastActiveDefinition'

Fields

lastModifiedOn :: Maybe POSIX

The date and time the blueprint was last modified.

blueprintServiceLocation :: Maybe Text

Specifies a path in Amazon S3 where the blueprint is copied when you create or update the blueprint.

parameterSpec :: Maybe Text

A JSON string specifying the parameters for the blueprint.

description :: Maybe Text

The description of the blueprint.

blueprintLocation :: Maybe Text

Specifies a path in Amazon S3 where the blueprint is published by the Glue developer.

newLastActiveDefinition :: LastActiveDefinition

Create a value of LastActiveDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedOn:LastActiveDefinition', lastActiveDefinition_lastModifiedOn - The date and time the blueprint was last modified.

$sel:blueprintServiceLocation:LastActiveDefinition', lastActiveDefinition_blueprintServiceLocation - Specifies a path in Amazon S3 where the blueprint is copied when you create or update the blueprint.

$sel:parameterSpec:LastActiveDefinition', lastActiveDefinition_parameterSpec - A JSON string specifying the parameters for the blueprint.

$sel:description:LastActiveDefinition', lastActiveDefinition_description - The description of the blueprint.

$sel:blueprintLocation:LastActiveDefinition', lastActiveDefinition_blueprintLocation - Specifies a path in Amazon S3 where the blueprint is published by the Glue developer.

lastActiveDefinition_lastModifiedOn :: Lens' LastActiveDefinition (Maybe UTCTime)

The date and time the blueprint was last modified.

lastActiveDefinition_blueprintServiceLocation :: Lens' LastActiveDefinition (Maybe Text)

Specifies a path in Amazon S3 where the blueprint is copied when you create or update the blueprint.

lastActiveDefinition_parameterSpec :: Lens' LastActiveDefinition (Maybe Text)

A JSON string specifying the parameters for the blueprint.

lastActiveDefinition_description :: Lens' LastActiveDefinition (Maybe Text)

The description of the blueprint.

lastActiveDefinition_blueprintLocation :: Lens' LastActiveDefinition (Maybe Text)

Specifies a path in Amazon S3 where the blueprint is published by the Glue developer.