Network.AWS.Glue.Types.LastCrawlInfo

Description

Documentation

data LastCrawlInfo

Status and error information about the most recent crawl.

See: newLastCrawlInfo smart constructor.

Constructors

LastCrawlInfo'

Fields

status :: Maybe LastCrawlStatus

Status of the last crawl.

logGroup :: Maybe Text

The log group for the last crawl.

messagePrefix :: Maybe Text

The prefix for a message about this crawl.

startTime :: Maybe POSIX

The time at which the crawl started.

errorMessage :: Maybe Text

If an error occurred, the error information about the last crawl.

logStream :: Maybe Text

The log stream for the last crawl.

newLastCrawlInfo :: LastCrawlInfo

Create a value of LastCrawlInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:LastCrawlInfo', lastCrawlInfo_status - Status of the last crawl.

$sel:logGroup:LastCrawlInfo', lastCrawlInfo_logGroup - The log group for the last crawl.

$sel:messagePrefix:LastCrawlInfo', lastCrawlInfo_messagePrefix - The prefix for a message about this crawl.

$sel:startTime:LastCrawlInfo', lastCrawlInfo_startTime - The time at which the crawl started.

$sel:errorMessage:LastCrawlInfo', lastCrawlInfo_errorMessage - If an error occurred, the error information about the last crawl.

$sel:logStream:LastCrawlInfo', lastCrawlInfo_logStream - The log stream for the last crawl.

lastCrawlInfo_status :: Lens' LastCrawlInfo (Maybe LastCrawlStatus)

Status of the last crawl.

lastCrawlInfo_logGroup :: Lens' LastCrawlInfo (Maybe Text)

The log group for the last crawl.

lastCrawlInfo_messagePrefix :: Lens' LastCrawlInfo (Maybe Text)

The prefix for a message about this crawl.

lastCrawlInfo_startTime :: Lens' LastCrawlInfo (Maybe UTCTime)

The time at which the crawl started.

lastCrawlInfo_errorMessage :: Lens' LastCrawlInfo (Maybe Text)

If an error occurred, the error information about the last crawl.

lastCrawlInfo_logStream :: Lens' LastCrawlInfo (Maybe Text)

The log stream for the last crawl.