Network.AWS.Glue.Types.MLUserDataEncryption

Description

Documentation

data MLUserDataEncryption

The encryption-at-rest settings of the transform that apply to accessing user data.

See: newMLUserDataEncryption smart constructor.

Constructors

MLUserDataEncryption'

Fields

kmsKeyId :: Maybe Text

The ID for the customer-provided KMS key.

mlUserDataEncryptionMode :: MLUserDataEncryptionModeString

The encryption mode applied to user data. Valid values are:

DISABLED: encryption is disabled

SSEKMS: use of server-side encryption with Key Management Service (SSE-KMS) for user data stored in Amazon S3.

newMLUserDataEncryption

Arguments

:: MLUserDataEncryptionModeString
$sel:mlUserDataEncryptionMode:MLUserDataEncryption'

-> MLUserDataEncryption

Create a value of MLUserDataEncryption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyId:MLUserDataEncryption', mLUserDataEncryption_kmsKeyId - The ID for the customer-provided KMS key.

$sel:mlUserDataEncryptionMode:MLUserDataEncryption', mLUserDataEncryption_mlUserDataEncryptionMode - The encryption mode applied to user data. Valid values are:

DISABLED: encryption is disabled

SSEKMS: use of server-side encryption with Key Management Service (SSE-KMS) for user data stored in Amazon S3.

mLUserDataEncryption_kmsKeyId :: Lens' MLUserDataEncryption (Maybe Text)

The ID for the customer-provided KMS key.

mLUserDataEncryption_mlUserDataEncryptionMode :: Lens' MLUserDataEncryption MLUserDataEncryptionModeString

The encryption mode applied to user data. Valid values are:

DISABLED: encryption is disabled

SSEKMS: use of server-side encryption with Key Management Service (SSE-KMS) for user data stored in Amazon S3.