Network.AWS.Glue.Types.Node

Description

Documentation

data Node

A node represents an Glue component (trigger, crawler, or job) on a workflow graph.

See: newNode smart constructor.

Constructors

Node'

Fields

jobDetails :: Maybe JobNodeDetails

Details of the Job when the node represents a Job.

triggerDetails :: Maybe TriggerNodeDetails

Details of the Trigger when the node represents a Trigger.

name :: Maybe Text

The name of the Glue component represented by the node.

crawlerDetails :: Maybe CrawlerNodeDetails

Details of the crawler when the node represents a crawler.

uniqueId :: Maybe Text

The unique Id assigned to the node within the workflow.

type' :: Maybe NodeType

The type of Glue component represented by the node.

newNode :: Node

Create a value of Node with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobDetails:Node', node_jobDetails - Details of the Job when the node represents a Job.

$sel:triggerDetails:Node', node_triggerDetails - Details of the Trigger when the node represents a Trigger.

$sel:name:Node', node_name - The name of the Glue component represented by the node.

$sel:crawlerDetails:Node', node_crawlerDetails - Details of the crawler when the node represents a crawler.

$sel:uniqueId:Node', node_uniqueId - The unique Id assigned to the node within the workflow.

$sel:type':Node', node_type - The type of Glue component represented by the node.

node_jobDetails :: Lens' Node (Maybe JobNodeDetails)

Details of the Job when the node represents a Job.

node_triggerDetails :: Lens' Node (Maybe TriggerNodeDetails)

Details of the Trigger when the node represents a Trigger.

node_name :: Lens' Node (Maybe Text)

The name of the Glue component represented by the node.

node_crawlerDetails :: Lens' Node (Maybe CrawlerNodeDetails)

Details of the crawler when the node represents a crawler.

node_uniqueId :: Lens' Node (Maybe Text)

The unique Id assigned to the node within the workflow.

node_type :: Lens' Node (Maybe NodeType)

The type of Glue component represented by the node.