Network.AWS.Glue.Types.PartitionInput

Description

Documentation

data PartitionInput

The structure used to create and update a partition.

See: newPartitionInput smart constructor.

Constructors

PartitionInput'

Fields

values :: Maybe [Text]

The values of the partition. Although this parameter is not required by the SDK, you must specify this parameter for a valid input.

The values for the keys for the new partition must be passed as an array of String objects that must be ordered in the same order as the partition keys appearing in the Amazon S3 prefix. Otherwise Glue will add the values to the wrong keys.

storageDescriptor :: Maybe StorageDescriptor

Provides information about the physical location where the partition is stored.

lastAnalyzedTime :: Maybe POSIX

The last time at which column statistics were computed for this partition.

lastAccessTime :: Maybe POSIX

The last time at which the partition was accessed.

parameters :: Maybe (HashMap Text Text)

These key-value pairs define partition parameters.

newPartitionInput :: PartitionInput

Create a value of PartitionInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:PartitionInput', partitionInput_values - The values of the partition. Although this parameter is not required by the SDK, you must specify this parameter for a valid input.

The values for the keys for the new partition must be passed as an array of String objects that must be ordered in the same order as the partition keys appearing in the Amazon S3 prefix. Otherwise Glue will add the values to the wrong keys.

$sel:storageDescriptor:PartitionInput', partitionInput_storageDescriptor - Provides information about the physical location where the partition is stored.

$sel:lastAnalyzedTime:PartitionInput', partitionInput_lastAnalyzedTime - The last time at which column statistics were computed for this partition.

$sel:lastAccessTime:PartitionInput', partitionInput_lastAccessTime - The last time at which the partition was accessed.

$sel:parameters:PartitionInput', partitionInput_parameters - These key-value pairs define partition parameters.

partitionInput_values :: Lens' PartitionInput (Maybe [Text])

The values of the partition. Although this parameter is not required by the SDK, you must specify this parameter for a valid input.

The values for the keys for the new partition must be passed as an array of String objects that must be ordered in the same order as the partition keys appearing in the Amazon S3 prefix. Otherwise Glue will add the values to the wrong keys.

partitionInput_storageDescriptor :: Lens' PartitionInput (Maybe StorageDescriptor)

Provides information about the physical location where the partition is stored.

partitionInput_lastAnalyzedTime :: Lens' PartitionInput (Maybe UTCTime)

The last time at which column statistics were computed for this partition.

partitionInput_lastAccessTime :: Lens' PartitionInput (Maybe UTCTime)

The last time at which the partition was accessed.

partitionInput_parameters :: Lens' PartitionInput (Maybe (HashMap Text Text))

These key-value pairs define partition parameters.