Network.AWS.Glue.Types.RecrawlPolicy

Description

Documentation

data RecrawlPolicy

When crawling an Amazon S3 data source after the first crawl is complete, specifies whether to crawl the entire dataset again or to crawl only folders that were added since the last crawler run. For more information, see Incremental Crawls in Glue in the developer guide.

See: newRecrawlPolicy smart constructor.

Constructors

RecrawlPolicy'

Fields

recrawlBehavior :: Maybe RecrawlBehavior

Specifies whether to crawl the entire dataset again or to crawl only folders that were added since the last crawler run.

A value of CRAWL_EVERYTHING specifies crawling the entire dataset again.

A value of CRAWL_NEW_FOLDERS_ONLY specifies crawling only folders that were added since the last crawler run.

newRecrawlPolicy :: RecrawlPolicy

Create a value of RecrawlPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recrawlBehavior:RecrawlPolicy', recrawlPolicy_recrawlBehavior - Specifies whether to crawl the entire dataset again or to crawl only folders that were added since the last crawler run.

A value of CRAWL_EVERYTHING specifies crawling the entire dataset again.

A value of CRAWL_NEW_FOLDERS_ONLY specifies crawling only folders that were added since the last crawler run.

recrawlPolicy_recrawlBehavior :: Lens' RecrawlPolicy (Maybe RecrawlBehavior)

Specifies whether to crawl the entire dataset again or to crawl only folders that were added since the last crawler run.

A value of CRAWL_EVERYTHING specifies crawling the entire dataset again.

A value of CRAWL_NEW_FOLDERS_ONLY specifies crawling only folders that were added since the last crawler run.