Network.AWS.Glue.Types.SchemaId

Description

Documentation

data SchemaId

The unique ID of the schema in the Glue schema registry.

See: newSchemaId smart constructor.

Constructors

SchemaId'

Fields

registryName :: Maybe Text

The name of the schema registry that contains the schema.

schemaName :: Maybe Text

The name of the schema. One of SchemaArn or SchemaName has to be provided.

schemaArn :: Maybe Text

The Amazon Resource Name (ARN) of the schema. One of SchemaArn or SchemaName has to be provided.

newSchemaId :: SchemaId

Create a value of SchemaId with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryName:SchemaId', schemaId_registryName - The name of the schema registry that contains the schema.

$sel:schemaName:SchemaId', schemaId_schemaName - The name of the schema. One of SchemaArn or SchemaName has to be provided.

$sel:schemaArn:SchemaId', schemaId_schemaArn - The Amazon Resource Name (ARN) of the schema. One of SchemaArn or SchemaName has to be provided.

schemaId_registryName :: Lens' SchemaId (Maybe Text)

The name of the schema registry that contains the schema.

schemaId_schemaName :: Lens' SchemaId (Maybe Text)

The name of the schema. One of SchemaArn or SchemaName has to be provided.

schemaId_schemaArn :: Lens' SchemaId (Maybe Text)

The Amazon Resource Name (ARN) of the schema. One of SchemaArn or SchemaName has to be provided.