Network.AWS.Glue.Types.SchemaListItem

Description

Documentation

data SchemaListItem

An object that contains minimal details for a schema.

See: newSchemaListItem smart constructor.

Constructors

SchemaListItem'

Fields

schemaArn :: Maybe Text

The Amazon Resource Name (ARN) for the schema.

updatedTime :: Maybe Text

The date and time that a schema was updated.

createdTime :: Maybe Text

The date and time that a schema was created.

registryName :: Maybe Text

the name of the registry where the schema resides.

description :: Maybe Text

A description for the schema.

schemaName :: Maybe Text

The name of the schema.

schemaStatus :: Maybe SchemaStatus

The status of the schema.

newSchemaListItem :: SchemaListItem

Create a value of SchemaListItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaArn:SchemaListItem', schemaListItem_schemaArn - The Amazon Resource Name (ARN) for the schema.

$sel:updatedTime:SchemaListItem', schemaListItem_updatedTime - The date and time that a schema was updated.

$sel:createdTime:SchemaListItem', schemaListItem_createdTime - The date and time that a schema was created.

$sel:registryName:SchemaListItem', schemaListItem_registryName - the name of the registry where the schema resides.

$sel:description:SchemaListItem', schemaListItem_description - A description for the schema.

$sel:schemaName:SchemaListItem', schemaListItem_schemaName - The name of the schema.

$sel:schemaStatus:SchemaListItem', schemaListItem_schemaStatus - The status of the schema.

schemaListItem_schemaArn :: Lens' SchemaListItem (Maybe Text)

The Amazon Resource Name (ARN) for the schema.

schemaListItem_updatedTime :: Lens' SchemaListItem (Maybe Text)

The date and time that a schema was updated.

schemaListItem_createdTime :: Lens' SchemaListItem (Maybe Text)

The date and time that a schema was created.

schemaListItem_registryName :: Lens' SchemaListItem (Maybe Text)

the name of the registry where the schema resides.

schemaListItem_description :: Lens' SchemaListItem (Maybe Text)

A description for the schema.

schemaListItem_schemaName :: Lens' SchemaListItem (Maybe Text)

The name of the schema.

schemaListItem_schemaStatus :: Lens' SchemaListItem (Maybe SchemaStatus)

The status of the schema.