Network.AWS.Glue.Types.SchemaVersionListItem

Description

Documentation

data SchemaVersionListItem

An object containing the details about a schema version.

See: newSchemaVersionListItem smart constructor.

Constructors

SchemaVersionListItem'

Fields

status :: Maybe SchemaVersionStatus

The status of the schema version.

schemaArn :: Maybe Text

The Amazon Resource Name (ARN) of the schema.

schemaVersionId :: Maybe Text

The unique identifier of the schema version.

createdTime :: Maybe Text

The date and time the schema version was created.

versionNumber :: Maybe Natural

The version number of the schema.

newSchemaVersionListItem :: SchemaVersionListItem

Create a value of SchemaVersionListItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SchemaVersionListItem', schemaVersionListItem_status - The status of the schema version.

$sel:schemaArn:SchemaVersionListItem', schemaVersionListItem_schemaArn - The Amazon Resource Name (ARN) of the schema.

$sel:schemaVersionId:SchemaVersionListItem', schemaVersionListItem_schemaVersionId - The unique identifier of the schema version.

$sel:createdTime:SchemaVersionListItem', schemaVersionListItem_createdTime - The date and time the schema version was created.

$sel:versionNumber:SchemaVersionListItem', schemaVersionListItem_versionNumber - The version number of the schema.

schemaVersionListItem_status :: Lens' SchemaVersionListItem (Maybe SchemaVersionStatus)

The status of the schema version.

schemaVersionListItem_schemaArn :: Lens' SchemaVersionListItem (Maybe Text)

The Amazon Resource Name (ARN) of the schema.

schemaVersionListItem_schemaVersionId :: Lens' SchemaVersionListItem (Maybe Text)

The unique identifier of the schema version.

schemaVersionListItem_createdTime :: Lens' SchemaVersionListItem (Maybe Text)

The date and time the schema version was created.

schemaVersionListItem_versionNumber :: Lens' SchemaVersionListItem (Maybe Natural)

The version number of the schema.