Network.AWS.Glue.Types.SkewedInfo

Description

Documentation

data SkewedInfo

Specifies skewed values in a table. Skewed values are those that occur with very high frequency.

See: newSkewedInfo smart constructor.

Constructors

SkewedInfo'

Fields

skewedColumnNames :: Maybe [Text]

A list of names of columns that contain skewed values.

skewedColumnValues :: Maybe [Text]

A list of values that appear so frequently as to be considered skewed.

skewedColumnValueLocationMaps :: Maybe (HashMap Text Text)

A mapping of skewed values to the columns that contain them.

newSkewedInfo :: SkewedInfo

Create a value of SkewedInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:skewedColumnNames:SkewedInfo', skewedInfo_skewedColumnNames - A list of names of columns that contain skewed values.

$sel:skewedColumnValues:SkewedInfo', skewedInfo_skewedColumnValues - A list of values that appear so frequently as to be considered skewed.

$sel:skewedColumnValueLocationMaps:SkewedInfo', skewedInfo_skewedColumnValueLocationMaps - A mapping of skewed values to the columns that contain them.

skewedInfo_skewedColumnNames :: Lens' SkewedInfo (Maybe [Text])

A list of names of columns that contain skewed values.

skewedInfo_skewedColumnValues :: Lens' SkewedInfo (Maybe [Text])

A list of values that appear so frequently as to be considered skewed.

skewedInfo_skewedColumnValueLocationMaps :: Lens' SkewedInfo (Maybe (HashMap Text Text))

A mapping of skewed values to the columns that contain them.