Network.AWS.Glue.Types.SortCriterion

Description

Documentation

data SortCriterion

Specifies a field to sort by and a sort order.

See: newSortCriterion smart constructor.

Constructors

SortCriterion'

Fields

sort :: Maybe Sort

An ascending or descending sort.

fieldName :: Maybe Text

The name of the field on which to sort.

newSortCriterion :: SortCriterion

Create a value of SortCriterion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sort:SortCriterion', sortCriterion_sort - An ascending or descending sort.

$sel:fieldName:SortCriterion', sortCriterion_fieldName - The name of the field on which to sort.

sortCriterion_sort :: Lens' SortCriterion (Maybe Sort)

An ascending or descending sort.

sortCriterion_fieldName :: Lens' SortCriterion (Maybe Text)

The name of the field on which to sort.