Network.AWS.Glue.Types.StorageDescriptor

Description

Documentation

data StorageDescriptor

Describes the physical storage of table data.

See: newStorageDescriptor smart constructor.

Constructors

StorageDescriptor'

Fields

compressed :: Maybe Bool

True if the data in the table is compressed, or False if not.

numberOfBuckets :: Maybe Int

Must be specified if the table contains any dimension columns.

skewedInfo :: Maybe SkewedInfo

The information about values that appear frequently in a column (skewed values).

schemaReference :: Maybe SchemaReference

An object that references a schema stored in the Glue Schema Registry.

When creating a table, you can pass an empty list of columns for the schema, and instead use a schema reference.

sortColumns :: Maybe [Order]

A list specifying the sort order of each bucket in the table.

outputFormat :: Maybe Text

The output format: SequenceFileOutputFormat (binary), or IgnoreKeyTextOutputFormat, or a custom format.

bucketColumns :: Maybe [Text]

A list of reducer grouping columns, clustering columns, and bucketing columns in the table.

serdeInfo :: Maybe SerDeInfo

The serialization/deserialization (SerDe) information.

location :: Maybe Text

The physical location of the table. By default, this takes the form of the warehouse location, followed by the database location in the warehouse, followed by the table name.

inputFormat :: Maybe Text

The input format: SequenceFileInputFormat (binary), or TextInputFormat, or a custom format.

columns :: Maybe [Column]

A list of the Columns in the table.

storedAsSubDirectories :: Maybe Bool

True if the table data is stored in subdirectories, or False if not.

parameters :: Maybe (HashMap Text Text)

The user-supplied properties in key-value form.

newStorageDescriptor :: StorageDescriptor

Create a value of StorageDescriptor with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:compressed:StorageDescriptor', storageDescriptor_compressed - True if the data in the table is compressed, or False if not.

$sel:numberOfBuckets:StorageDescriptor', storageDescriptor_numberOfBuckets - Must be specified if the table contains any dimension columns.

$sel:skewedInfo:StorageDescriptor', storageDescriptor_skewedInfo - The information about values that appear frequently in a column (skewed values).

$sel:schemaReference:StorageDescriptor', storageDescriptor_schemaReference - An object that references a schema stored in the Glue Schema Registry.

When creating a table, you can pass an empty list of columns for the schema, and instead use a schema reference.

$sel:sortColumns:StorageDescriptor', storageDescriptor_sortColumns - A list specifying the sort order of each bucket in the table.

$sel:outputFormat:StorageDescriptor', storageDescriptor_outputFormat - The output format: SequenceFileOutputFormat (binary), or IgnoreKeyTextOutputFormat, or a custom format.

$sel:bucketColumns:StorageDescriptor', storageDescriptor_bucketColumns - A list of reducer grouping columns, clustering columns, and bucketing columns in the table.

$sel:serdeInfo:StorageDescriptor', storageDescriptor_serdeInfo - The serialization/deserialization (SerDe) information.

$sel:location:StorageDescriptor', storageDescriptor_location - The physical location of the table. By default, this takes the form of the warehouse location, followed by the database location in the warehouse, followed by the table name.

$sel:inputFormat:StorageDescriptor', storageDescriptor_inputFormat - The input format: SequenceFileInputFormat (binary), or TextInputFormat, or a custom format.

$sel:columns:StorageDescriptor', storageDescriptor_columns - A list of the Columns in the table.

$sel:storedAsSubDirectories:StorageDescriptor', storageDescriptor_storedAsSubDirectories - True if the table data is stored in subdirectories, or False if not.

$sel:parameters:StorageDescriptor', storageDescriptor_parameters - The user-supplied properties in key-value form.

storageDescriptor_compressed :: Lens' StorageDescriptor (Maybe Bool)

True if the data in the table is compressed, or False if not.

storageDescriptor_numberOfBuckets :: Lens' StorageDescriptor (Maybe Int)

Must be specified if the table contains any dimension columns.

storageDescriptor_skewedInfo :: Lens' StorageDescriptor (Maybe SkewedInfo)

The information about values that appear frequently in a column (skewed values).

storageDescriptor_schemaReference :: Lens' StorageDescriptor (Maybe SchemaReference)

An object that references a schema stored in the Glue Schema Registry.

When creating a table, you can pass an empty list of columns for the schema, and instead use a schema reference.

storageDescriptor_sortColumns :: Lens' StorageDescriptor (Maybe [Order])

A list specifying the sort order of each bucket in the table.

storageDescriptor_outputFormat :: Lens' StorageDescriptor (Maybe Text)

The output format: SequenceFileOutputFormat (binary), or IgnoreKeyTextOutputFormat, or a custom format.

storageDescriptor_bucketColumns :: Lens' StorageDescriptor (Maybe [Text])

A list of reducer grouping columns, clustering columns, and bucketing columns in the table.

storageDescriptor_serdeInfo :: Lens' StorageDescriptor (Maybe SerDeInfo)

The serialization/deserialization (SerDe) information.

storageDescriptor_location :: Lens' StorageDescriptor (Maybe Text)

The physical location of the table. By default, this takes the form of the warehouse location, followed by the database location in the warehouse, followed by the table name.

storageDescriptor_inputFormat :: Lens' StorageDescriptor (Maybe Text)

The input format: SequenceFileInputFormat (binary), or TextInputFormat, or a custom format.

storageDescriptor_columns :: Lens' StorageDescriptor (Maybe [Column])

A list of the Columns in the table.

storageDescriptor_storedAsSubDirectories :: Lens' StorageDescriptor (Maybe Bool)

True if the table data is stored in subdirectories, or False if not.

storageDescriptor_parameters :: Lens' StorageDescriptor (Maybe (HashMap Text Text))

The user-supplied properties in key-value form.