Network.AWS.Glue.Types.Table

Description

Documentation

data Table

Represents a collection of related data organized in columns and rows.

See: newTable smart constructor.

Constructors

Table'

Fields

viewOriginalText :: Maybe Text

If the table is a view, the original text of the view; otherwise null.

tableType :: Maybe Text

The type of this table (EXTERNAL_TABLE, VIRTUAL_VIEW, etc.).

catalogId :: Maybe Text

The ID of the Data Catalog in which the table resides.

storageDescriptor :: Maybe StorageDescriptor

A storage descriptor containing information about the physical storage of this table.

lastAnalyzedTime :: Maybe POSIX

The last time that column statistics were computed for this table.

viewExpandedText :: Maybe Text

If the table is a view, the expanded text of the view; otherwise null.

targetTable :: Maybe TableIdentifier

A TableIdentifier structure that describes a target table for resource linking.

retention :: Maybe Natural

The retention time for this table.

updateTime :: Maybe POSIX

The last time that the table was updated.

createTime :: Maybe POSIX

The time when the table definition was created in the Data Catalog.

owner :: Maybe Text

The owner of the table.

description :: Maybe Text

A description of the table.

partitionKeys :: Maybe [Column]

A list of columns by which the table is partitioned. Only primitive types are supported as partition keys.

When you create a table used by Amazon Athena, and you do not specify any partitionKeys, you must at least set the value of partitionKeys to an empty list. For example:

"PartitionKeys": []

createdBy :: Maybe Text

The person or entity who created the table.

lastAccessTime :: Maybe POSIX

The last time that the table was accessed. This is usually taken from HDFS, and might not be reliable.

isRegisteredWithLakeFormation :: Maybe Bool

Indicates whether the table has been registered with Lake Formation.

parameters :: Maybe (HashMap Text Text)

These key-value pairs define properties associated with the table.

databaseName :: Maybe Text

The name of the database where the table metadata resides. For Hive compatibility, this must be all lowercase.

name :: Text

The table name. For Hive compatibility, this must be entirely lowercase.

newTable

Arguments

:: Text
$sel:name:Table'

-> Table

Create a value of Table with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:viewOriginalText:Table', table_viewOriginalText - If the table is a view, the original text of the view; otherwise null.

$sel:tableType:Table', table_tableType - The type of this table (EXTERNAL_TABLE, VIRTUAL_VIEW, etc.).

$sel:catalogId:Table', table_catalogId - The ID of the Data Catalog in which the table resides.

$sel:storageDescriptor:Table', table_storageDescriptor - A storage descriptor containing information about the physical storage of this table.

$sel:lastAnalyzedTime:Table', table_lastAnalyzedTime - The last time that column statistics were computed for this table.

$sel:viewExpandedText:Table', table_viewExpandedText - If the table is a view, the expanded text of the view; otherwise null.

$sel:targetTable:Table', table_targetTable - A TableIdentifier structure that describes a target table for resource linking.

$sel:retention:Table', table_retention - The retention time for this table.

$sel:updateTime:Table', table_updateTime - The last time that the table was updated.

$sel:createTime:Table', table_createTime - The time when the table definition was created in the Data Catalog.

$sel:owner:Table', table_owner - The owner of the table.

$sel:description:Table', table_description - A description of the table.

$sel:partitionKeys:Table', table_partitionKeys - A list of columns by which the table is partitioned. Only primitive types are supported as partition keys.

When you create a table used by Amazon Athena, and you do not specify any partitionKeys, you must at least set the value of partitionKeys to an empty list. For example:

"PartitionKeys": []

$sel:createdBy:Table', table_createdBy - The person or entity who created the table.

$sel:lastAccessTime:Table', table_lastAccessTime - The last time that the table was accessed. This is usually taken from HDFS, and might not be reliable.

$sel:isRegisteredWithLakeFormation:Table', table_isRegisteredWithLakeFormation - Indicates whether the table has been registered with Lake Formation.

$sel:parameters:Table', table_parameters - These key-value pairs define properties associated with the table.

$sel:databaseName:Table', table_databaseName - The name of the database where the table metadata resides. For Hive compatibility, this must be all lowercase.

$sel:name:Table', table_name - The table name. For Hive compatibility, this must be entirely lowercase.

table_viewOriginalText :: Lens' Table (Maybe Text)

If the table is a view, the original text of the view; otherwise null.

table_tableType :: Lens' Table (Maybe Text)

The type of this table (EXTERNAL_TABLE, VIRTUAL_VIEW, etc.).

table_catalogId :: Lens' Table (Maybe Text)

The ID of the Data Catalog in which the table resides.

table_storageDescriptor :: Lens' Table (Maybe StorageDescriptor)

A storage descriptor containing information about the physical storage of this table.

table_lastAnalyzedTime :: Lens' Table (Maybe UTCTime)

The last time that column statistics were computed for this table.

table_viewExpandedText :: Lens' Table (Maybe Text)

If the table is a view, the expanded text of the view; otherwise null.

table_targetTable :: Lens' Table (Maybe TableIdentifier)

A TableIdentifier structure that describes a target table for resource linking.

table_retention :: Lens' Table (Maybe Natural)

The retention time for this table.

table_updateTime :: Lens' Table (Maybe UTCTime)

The last time that the table was updated.

table_createTime :: Lens' Table (Maybe UTCTime)

The time when the table definition was created in the Data Catalog.

table_owner :: Lens' Table (Maybe Text)

The owner of the table.

table_description :: Lens' Table (Maybe Text)

A description of the table.

table_partitionKeys :: Lens' Table (Maybe [Column])

A list of columns by which the table is partitioned. Only primitive types are supported as partition keys.

When you create a table used by Amazon Athena, and you do not specify any partitionKeys, you must at least set the value of partitionKeys to an empty list. For example:

"PartitionKeys": []

table_createdBy :: Lens' Table (Maybe Text)

The person or entity who created the table.

table_lastAccessTime :: Lens' Table (Maybe UTCTime)

The last time that the table was accessed. This is usually taken from HDFS, and might not be reliable.

table_isRegisteredWithLakeFormation :: Lens' Table (Maybe Bool)

Indicates whether the table has been registered with Lake Formation.

table_parameters :: Lens' Table (Maybe (HashMap Text Text))

These key-value pairs define properties associated with the table.

table_databaseName :: Lens' Table (Maybe Text)

The name of the database where the table metadata resides. For Hive compatibility, this must be all lowercase.

table_name :: Lens' Table Text

The table name. For Hive compatibility, this must be entirely lowercase.