Network.AWS.Glue.Types.TaskRun

Description

Documentation

data TaskRun

The sampling parameters that are associated with the machine learning transform.

See: newTaskRun smart constructor.

Constructors

TaskRun'

Fields

completedOn :: Maybe POSIX

The last point in time that the requested task run was completed.

status :: Maybe TaskStatusType

The current status of the requested task run.

lastModifiedOn :: Maybe POSIX

The last point in time that the requested task run was updated.

errorString :: Maybe Text

The list of error strings associated with this task run.

startedOn :: Maybe POSIX

The date and time that this task run started.

logGroupName :: Maybe Text

The names of the log group for secure logging, associated with this task run.

executionTime :: Maybe Int

The amount of time (in seconds) that the task run consumed resources.

properties :: Maybe TaskRunProperties

Specifies configuration properties associated with this task run.

transformId :: Maybe Text

The unique identifier for the transform.

taskRunId :: Maybe Text

The unique identifier for this task run.

newTaskRun :: TaskRun

Create a value of TaskRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:completedOn:TaskRun', taskRun_completedOn - The last point in time that the requested task run was completed.

$sel:status:TaskRun', taskRun_status - The current status of the requested task run.

$sel:lastModifiedOn:TaskRun', taskRun_lastModifiedOn - The last point in time that the requested task run was updated.

$sel:errorString:TaskRun', taskRun_errorString - The list of error strings associated with this task run.

$sel:startedOn:TaskRun', taskRun_startedOn - The date and time that this task run started.

$sel:logGroupName:TaskRun', taskRun_logGroupName - The names of the log group for secure logging, associated with this task run.

$sel:executionTime:TaskRun', taskRun_executionTime - The amount of time (in seconds) that the task run consumed resources.

$sel:properties:TaskRun', taskRun_properties - Specifies configuration properties associated with this task run.

$sel:transformId:TaskRun', taskRun_transformId - The unique identifier for the transform.

$sel:taskRunId:TaskRun', taskRun_taskRunId - The unique identifier for this task run.

taskRun_completedOn :: Lens' TaskRun (Maybe UTCTime)

The last point in time that the requested task run was completed.

taskRun_status :: Lens' TaskRun (Maybe TaskStatusType)

The current status of the requested task run.

taskRun_lastModifiedOn :: Lens' TaskRun (Maybe UTCTime)

The last point in time that the requested task run was updated.

taskRun_errorString :: Lens' TaskRun (Maybe Text)

The list of error strings associated with this task run.

taskRun_startedOn :: Lens' TaskRun (Maybe UTCTime)

The date and time that this task run started.

taskRun_logGroupName :: Lens' TaskRun (Maybe Text)

The names of the log group for secure logging, associated with this task run.

taskRun_executionTime :: Lens' TaskRun (Maybe Int)

The amount of time (in seconds) that the task run consumed resources.

taskRun_properties :: Lens' TaskRun (Maybe TaskRunProperties)

Specifies configuration properties associated with this task run.

taskRun_transformId :: Lens' TaskRun (Maybe Text)

The unique identifier for the transform.

taskRun_taskRunId :: Lens' TaskRun (Maybe Text)

The unique identifier for this task run.