Network.AWS.Glue.Types.TaskRunFilterCriteria

Description

Documentation

data TaskRunFilterCriteria

The criteria that are used to filter the task runs for the machine learning transform.

See: newTaskRunFilterCriteria smart constructor.

Constructors

TaskRunFilterCriteria'

Fields

status :: Maybe TaskStatusType

The current status of the task run.

startedAfter :: Maybe POSIX

Filter on task runs started after this date.

startedBefore :: Maybe POSIX

Filter on task runs started before this date.

taskRunType :: Maybe TaskType

The type of task run.

newTaskRunFilterCriteria :: TaskRunFilterCriteria

Create a value of TaskRunFilterCriteria with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:TaskRunFilterCriteria', taskRunFilterCriteria_status - The current status of the task run.

$sel:startedAfter:TaskRunFilterCriteria', taskRunFilterCriteria_startedAfter - Filter on task runs started after this date.

$sel:startedBefore:TaskRunFilterCriteria', taskRunFilterCriteria_startedBefore - Filter on task runs started before this date.

$sel:taskRunType:TaskRunFilterCriteria', taskRunFilterCriteria_taskRunType - The type of task run.

taskRunFilterCriteria_status :: Lens' TaskRunFilterCriteria (Maybe TaskStatusType)

The current status of the task run.

taskRunFilterCriteria_startedAfter :: Lens' TaskRunFilterCriteria (Maybe UTCTime)

Filter on task runs started after this date.

taskRunFilterCriteria_startedBefore :: Lens' TaskRunFilterCriteria (Maybe UTCTime)

Filter on task runs started before this date.

taskRunFilterCriteria_taskRunType :: Lens' TaskRunFilterCriteria (Maybe TaskType)

The type of task run.