Network.AWS.Glue.Types.TaskRunProperties

Description

Documentation

data TaskRunProperties

The configuration properties for the task run.

See: newTaskRunProperties smart constructor.

Constructors

TaskRunProperties'

Fields

taskType :: Maybe TaskType

The type of task run.

exportLabelsTaskRunProperties :: Maybe ExportLabelsTaskRunProperties

The configuration properties for an exporting labels task run.

labelingSetGenerationTaskRunProperties :: Maybe LabelingSetGenerationTaskRunProperties

The configuration properties for a labeling set generation task run.

findMatchesTaskRunProperties :: Maybe FindMatchesTaskRunProperties

The configuration properties for a find matches task run.

importLabelsTaskRunProperties :: Maybe ImportLabelsTaskRunProperties

The configuration properties for an importing labels task run.

newTaskRunProperties :: TaskRunProperties

Create a value of TaskRunProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskType:TaskRunProperties', taskRunProperties_taskType - The type of task run.

$sel:exportLabelsTaskRunProperties:TaskRunProperties', taskRunProperties_exportLabelsTaskRunProperties - The configuration properties for an exporting labels task run.

$sel:labelingSetGenerationTaskRunProperties:TaskRunProperties', taskRunProperties_labelingSetGenerationTaskRunProperties - The configuration properties for a labeling set generation task run.

$sel:findMatchesTaskRunProperties:TaskRunProperties', taskRunProperties_findMatchesTaskRunProperties - The configuration properties for a find matches task run.

$sel:importLabelsTaskRunProperties:TaskRunProperties', taskRunProperties_importLabelsTaskRunProperties - The configuration properties for an importing labels task run.

taskRunProperties_taskType :: Lens' TaskRunProperties (Maybe TaskType)

The type of task run.

taskRunProperties_exportLabelsTaskRunProperties :: Lens' TaskRunProperties (Maybe ExportLabelsTaskRunProperties)

The configuration properties for an exporting labels task run.

taskRunProperties_labelingSetGenerationTaskRunProperties :: Lens' TaskRunProperties (Maybe LabelingSetGenerationTaskRunProperties)

The configuration properties for a labeling set generation task run.

taskRunProperties_findMatchesTaskRunProperties :: Lens' TaskRunProperties (Maybe FindMatchesTaskRunProperties)

The configuration properties for a find matches task run.

taskRunProperties_importLabelsTaskRunProperties :: Lens' TaskRunProperties (Maybe ImportLabelsTaskRunProperties)

The configuration properties for an importing labels task run.