Network.AWS.Glue.Types.TaskRunSortCriteria

Description

Documentation

data TaskRunSortCriteria

The sorting criteria that are used to sort the list of task runs for the machine learning transform.

See: newTaskRunSortCriteria smart constructor.

Constructors

TaskRunSortCriteria'

Fields

column :: TaskRunSortColumnType

The column to be used to sort the list of task runs for the machine learning transform.

sortDirection :: SortDirectionType

The sort direction to be used to sort the list of task runs for the machine learning transform.

newTaskRunSortCriteria

Arguments

:: TaskRunSortColumnType
$sel:column:TaskRunSortCriteria'

-> SortDirectionType
$sel:sortDirection:TaskRunSortCriteria'

-> TaskRunSortCriteria

Create a value of TaskRunSortCriteria with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:column:TaskRunSortCriteria', taskRunSortCriteria_column - The column to be used to sort the list of task runs for the machine learning transform.

$sel:sortDirection:TaskRunSortCriteria', taskRunSortCriteria_sortDirection - The sort direction to be used to sort the list of task runs for the machine learning transform.

taskRunSortCriteria_column :: Lens' TaskRunSortCriteria TaskRunSortColumnType

The column to be used to sort the list of task runs for the machine learning transform.

taskRunSortCriteria_sortDirection :: Lens' TaskRunSortCriteria SortDirectionType

The sort direction to be used to sort the list of task runs for the machine learning transform.