Network.AWS.Glue.Types.Trigger

Description

Documentation

data Trigger

Information about a specific trigger.

See: newTrigger smart constructor.

Constructors

Trigger'

Fields

workflowName :: Maybe Text

The name of the workflow associated with the trigger.

id :: Maybe Text

Reserved for future use.

actions :: Maybe [Action]

The actions initiated by this trigger.

name :: Maybe Text

The name of the trigger.

state :: Maybe TriggerState

The current state of the trigger.

predicate :: Maybe Predicate

The predicate of this trigger, which defines when it will fire.

description :: Maybe Text

A description of this trigger.

eventBatchingCondition :: Maybe EventBatchingCondition

Batch condition that must be met (specified number of events received or batch time window expired) before EventBridge event trigger fires.

type' :: Maybe TriggerType

The type of trigger that this is.

schedule :: Maybe Text

A cron expression used to specify the schedule (see Time-Based Schedules for Jobs and Crawlers. For example, to run something every day at 12:15 UTC, you would specify: cron(15 12 * * ? *).

newTrigger :: Trigger

Create a value of Trigger with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workflowName:Trigger', trigger_workflowName - The name of the workflow associated with the trigger.

$sel:id:Trigger', trigger_id - Reserved for future use.

$sel:actions:Trigger', trigger_actions - The actions initiated by this trigger.

$sel:name:Trigger', trigger_name - The name of the trigger.

$sel:state:Trigger', trigger_state - The current state of the trigger.

$sel:predicate:Trigger', trigger_predicate - The predicate of this trigger, which defines when it will fire.

$sel:description:Trigger', trigger_description - A description of this trigger.

$sel:eventBatchingCondition:Trigger', trigger_eventBatchingCondition - Batch condition that must be met (specified number of events received or batch time window expired) before EventBridge event trigger fires.

$sel:type':Trigger', trigger_type - The type of trigger that this is.

$sel:schedule:Trigger', trigger_schedule - A cron expression used to specify the schedule (see Time-Based Schedules for Jobs and Crawlers. For example, to run something every day at 12:15 UTC, you would specify: cron(15 12 * * ? *).

trigger_workflowName :: Lens' Trigger (Maybe Text)

The name of the workflow associated with the trigger.

trigger_id :: Lens' Trigger (Maybe Text)

Reserved for future use.

trigger_actions :: Lens' Trigger (Maybe [Action])

The actions initiated by this trigger.

trigger_name :: Lens' Trigger (Maybe Text)

The name of the trigger.

trigger_state :: Lens' Trigger (Maybe TriggerState)

The current state of the trigger.

trigger_predicate :: Lens' Trigger (Maybe Predicate)

The predicate of this trigger, which defines when it will fire.

trigger_description :: Lens' Trigger (Maybe Text)

A description of this trigger.

trigger_eventBatchingCondition :: Lens' Trigger (Maybe EventBatchingCondition)

Batch condition that must be met (specified number of events received or batch time window expired) before EventBridge event trigger fires.

trigger_type :: Lens' Trigger (Maybe TriggerType)

The type of trigger that this is.

trigger_schedule :: Lens' Trigger (Maybe Text)

A cron expression used to specify the schedule (see Time-Based Schedules for Jobs and Crawlers. For example, to run something every day at 12:15 UTC, you would specify: cron(15 12 * * ? *).