Network.AWS.Glue.Types.UpdateCsvClassifierRequest

Description

Documentation

data UpdateCsvClassifierRequest

Specifies a custom CSV classifier to be updated.

See: newUpdateCsvClassifierRequest smart constructor.

Constructors

UpdateCsvClassifierRequest'

Fields

containsHeader :: Maybe CsvHeaderOption

Indicates whether the CSV file contains a header.

delimiter :: Maybe Text

A custom symbol to denote what separates each column entry in the row.

disableValueTrimming :: Maybe Bool

Specifies not to trim values before identifying the type of column values. The default value is true.

header :: Maybe [Text]

A list of strings representing column names.

quoteSymbol :: Maybe Text

A custom symbol to denote what combines content into a single column value. It must be different from the column delimiter.

allowSingleColumn :: Maybe Bool

Enables the processing of files that contain only one column.

name :: Text

The name of the classifier.

newUpdateCsvClassifierRequest

Arguments

:: Text
$sel:name:UpdateCsvClassifierRequest'

-> UpdateCsvClassifierRequest

Create a value of UpdateCsvClassifierRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containsHeader:UpdateCsvClassifierRequest', updateCsvClassifierRequest_containsHeader - Indicates whether the CSV file contains a header.

$sel:delimiter:UpdateCsvClassifierRequest', updateCsvClassifierRequest_delimiter - A custom symbol to denote what separates each column entry in the row.

$sel:disableValueTrimming:UpdateCsvClassifierRequest', updateCsvClassifierRequest_disableValueTrimming - Specifies not to trim values before identifying the type of column values. The default value is true.

$sel:header:UpdateCsvClassifierRequest', updateCsvClassifierRequest_header - A list of strings representing column names.

$sel:quoteSymbol:UpdateCsvClassifierRequest', updateCsvClassifierRequest_quoteSymbol - A custom symbol to denote what combines content into a single column value. It must be different from the column delimiter.

$sel:allowSingleColumn:UpdateCsvClassifierRequest', updateCsvClassifierRequest_allowSingleColumn - Enables the processing of files that contain only one column.

$sel:name:UpdateCsvClassifierRequest', updateCsvClassifierRequest_name - The name of the classifier.

updateCsvClassifierRequest_containsHeader :: Lens' UpdateCsvClassifierRequest (Maybe CsvHeaderOption)

Indicates whether the CSV file contains a header.

updateCsvClassifierRequest_delimiter :: Lens' UpdateCsvClassifierRequest (Maybe Text)

A custom symbol to denote what separates each column entry in the row.

updateCsvClassifierRequest_disableValueTrimming :: Lens' UpdateCsvClassifierRequest (Maybe Bool)

Specifies not to trim values before identifying the type of column values. The default value is true.

updateCsvClassifierRequest_header :: Lens' UpdateCsvClassifierRequest (Maybe [Text])

A list of strings representing column names.

updateCsvClassifierRequest_quoteSymbol :: Lens' UpdateCsvClassifierRequest (Maybe Text)

A custom symbol to denote what combines content into a single column value. It must be different from the column delimiter.

updateCsvClassifierRequest_allowSingleColumn :: Lens' UpdateCsvClassifierRequest (Maybe Bool)

Enables the processing of files that contain only one column.

updateCsvClassifierRequest_name :: Lens' UpdateCsvClassifierRequest Text

The name of the classifier.