Network.AWS.Glue.Types.Workflow

Description

Documentation

data Workflow

A workflow is a collection of multiple dependent Glue jobs and crawlers that are run to complete a complex ETL task. A workflow manages the execution and monitoring of all its jobs and crawlers.

See: newWorkflow smart constructor.

Constructors

Workflow'

Fields

createdOn :: Maybe POSIX

The date and time when the workflow was created.

defaultRunProperties :: Maybe (HashMap Text Text)

A collection of properties to be used as part of each execution of the workflow. The run properties are made available to each job in the workflow. A job can modify the properties for the next jobs in the flow.

lastRun :: Maybe WorkflowRun

The information about the last execution of the workflow.

maxConcurrentRuns :: Maybe Int

You can use this parameter to prevent unwanted multiple updates to data, to control costs, or in some cases, to prevent exceeding the maximum number of concurrent runs of any of the component jobs. If you leave this parameter blank, there is no limit to the number of concurrent workflow runs.

lastModifiedOn :: Maybe POSIX

The date and time when the workflow was last modified.

name :: Maybe Text

The name of the workflow.

graph :: Maybe WorkflowGraph

The graph representing all the Glue components that belong to the workflow as nodes and directed connections between them as edges.

description :: Maybe Text

A description of the workflow.

blueprintDetails :: Maybe BlueprintDetails

This structure indicates the details of the blueprint that this particular workflow is created from.

newWorkflow :: Workflow

Create a value of Workflow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdOn:Workflow', workflow_createdOn - The date and time when the workflow was created.

$sel:defaultRunProperties:Workflow', workflow_defaultRunProperties - A collection of properties to be used as part of each execution of the workflow. The run properties are made available to each job in the workflow. A job can modify the properties for the next jobs in the flow.

$sel:lastRun:Workflow', workflow_lastRun - The information about the last execution of the workflow.

$sel:maxConcurrentRuns:Workflow', workflow_maxConcurrentRuns - You can use this parameter to prevent unwanted multiple updates to data, to control costs, or in some cases, to prevent exceeding the maximum number of concurrent runs of any of the component jobs. If you leave this parameter blank, there is no limit to the number of concurrent workflow runs.

$sel:lastModifiedOn:Workflow', workflow_lastModifiedOn - The date and time when the workflow was last modified.

$sel:name:Workflow', workflow_name - The name of the workflow.

$sel:graph:Workflow', workflow_graph - The graph representing all the Glue components that belong to the workflow as nodes and directed connections between them as edges.

$sel:description:Workflow', workflow_description - A description of the workflow.

$sel:blueprintDetails:Workflow', workflow_blueprintDetails - This structure indicates the details of the blueprint that this particular workflow is created from.

workflow_createdOn :: Lens' Workflow (Maybe UTCTime)

The date and time when the workflow was created.

workflow_defaultRunProperties :: Lens' Workflow (Maybe (HashMap Text Text))

A collection of properties to be used as part of each execution of the workflow. The run properties are made available to each job in the workflow. A job can modify the properties for the next jobs in the flow.

workflow_lastRun :: Lens' Workflow (Maybe WorkflowRun)

The information about the last execution of the workflow.

workflow_maxConcurrentRuns :: Lens' Workflow (Maybe Int)

You can use this parameter to prevent unwanted multiple updates to data, to control costs, or in some cases, to prevent exceeding the maximum number of concurrent runs of any of the component jobs. If you leave this parameter blank, there is no limit to the number of concurrent workflow runs.

workflow_lastModifiedOn :: Lens' Workflow (Maybe UTCTime)

The date and time when the workflow was last modified.

workflow_name :: Lens' Workflow (Maybe Text)

The name of the workflow.

workflow_graph :: Lens' Workflow (Maybe WorkflowGraph)

The graph representing all the Glue components that belong to the workflow as nodes and directed connections between them as edges.

workflow_description :: Lens' Workflow (Maybe Text)

A description of the workflow.

workflow_blueprintDetails :: Lens' Workflow (Maybe BlueprintDetails)

This structure indicates the details of the blueprint that this particular workflow is created from.