Network.AWS.Glue.Types.WorkflowRun

Description

Documentation

data WorkflowRun

A workflow run is an execution of a workflow providing all the runtime information.

See: newWorkflowRun smart constructor.

Constructors

WorkflowRun'

Fields

workflowRunId :: Maybe Text

The ID of this workflow run.

status :: Maybe WorkflowRunStatus

The status of the workflow run.

startingEventBatchCondition :: Maybe StartingEventBatchCondition

The batch condition that started the workflow run.

workflowRunProperties :: Maybe (HashMap Text Text)

The workflow run properties which were set during the run.

statistics :: Maybe WorkflowRunStatistics

The statistics of the run.

name :: Maybe Text

Name of the workflow that was run.

graph :: Maybe WorkflowGraph

The graph representing all the Glue components that belong to the workflow as nodes and directed connections between them as edges.

completedOn :: Maybe POSIX

The date and time when the workflow run completed.

errorMessage :: Maybe Text

This error message describes any error that may have occurred in starting the workflow run. Currently the only error message is "Concurrent runs exceeded for workflow: foo."

startedOn :: Maybe POSIX

The date and time when the workflow run was started.

previousRunId :: Maybe Text

The ID of the previous workflow run.

newWorkflowRun :: WorkflowRun

Create a value of WorkflowRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workflowRunId:WorkflowRun', workflowRun_workflowRunId - The ID of this workflow run.

$sel:status:WorkflowRun', workflowRun_status - The status of the workflow run.

$sel:startingEventBatchCondition:WorkflowRun', workflowRun_startingEventBatchCondition - The batch condition that started the workflow run.

$sel:workflowRunProperties:WorkflowRun', workflowRun_workflowRunProperties - The workflow run properties which were set during the run.

$sel:statistics:WorkflowRun', workflowRun_statistics - The statistics of the run.

$sel:name:WorkflowRun', workflowRun_name - Name of the workflow that was run.

$sel:graph:WorkflowRun', workflowRun_graph - The graph representing all the Glue components that belong to the workflow as nodes and directed connections between them as edges.

$sel:completedOn:WorkflowRun', workflowRun_completedOn - The date and time when the workflow run completed.

$sel:errorMessage:WorkflowRun', workflowRun_errorMessage - This error message describes any error that may have occurred in starting the workflow run. Currently the only error message is "Concurrent runs exceeded for workflow: foo."

$sel:startedOn:WorkflowRun', workflowRun_startedOn - The date and time when the workflow run was started.

$sel:previousRunId:WorkflowRun', workflowRun_previousRunId - The ID of the previous workflow run.

workflowRun_workflowRunId :: Lens' WorkflowRun (Maybe Text)

The ID of this workflow run.

workflowRun_status :: Lens' WorkflowRun (Maybe WorkflowRunStatus)

The status of the workflow run.

workflowRun_startingEventBatchCondition :: Lens' WorkflowRun (Maybe StartingEventBatchCondition)

The batch condition that started the workflow run.

workflowRun_workflowRunProperties :: Lens' WorkflowRun (Maybe (HashMap Text Text))

The workflow run properties which were set during the run.

workflowRun_statistics :: Lens' WorkflowRun (Maybe WorkflowRunStatistics)

The statistics of the run.

workflowRun_name :: Lens' WorkflowRun (Maybe Text)

Name of the workflow that was run.

workflowRun_graph :: Lens' WorkflowRun (Maybe WorkflowGraph)

The graph representing all the Glue components that belong to the workflow as nodes and directed connections between them as edges.

workflowRun_completedOn :: Lens' WorkflowRun (Maybe UTCTime)

The date and time when the workflow run completed.

workflowRun_errorMessage :: Lens' WorkflowRun (Maybe Text)

This error message describes any error that may have occurred in starting the workflow run. Currently the only error message is "Concurrent runs exceeded for workflow: foo."

workflowRun_startedOn :: Lens' WorkflowRun (Maybe UTCTime)

The date and time when the workflow run was started.

workflowRun_previousRunId :: Lens' WorkflowRun (Maybe Text)

The ID of the previous workflow run.