Network.AWS.Glue.Types.WorkflowRunStatistics

Description

Documentation

data WorkflowRunStatistics

Workflow run statistics provides statistics about the workflow run.

See: newWorkflowRunStatistics smart constructor.

Constructors

WorkflowRunStatistics'

Fields

runningActions :: Maybe Int

Total number Actions in running state.

stoppedActions :: Maybe Int

Total number of Actions that have stopped.

totalActions :: Maybe Int

Total number of Actions in the workflow run.

failedActions :: Maybe Int

Total number of Actions that have failed.

timeoutActions :: Maybe Int

Total number of Actions that timed out.

succeededActions :: Maybe Int

Total number of Actions that have succeeded.

newWorkflowRunStatistics :: WorkflowRunStatistics

Create a value of WorkflowRunStatistics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:runningActions:WorkflowRunStatistics', workflowRunStatistics_runningActions - Total number Actions in running state.

$sel:stoppedActions:WorkflowRunStatistics', workflowRunStatistics_stoppedActions - Total number of Actions that have stopped.

$sel:totalActions:WorkflowRunStatistics', workflowRunStatistics_totalActions - Total number of Actions in the workflow run.

$sel:failedActions:WorkflowRunStatistics', workflowRunStatistics_failedActions - Total number of Actions that have failed.

$sel:timeoutActions:WorkflowRunStatistics', workflowRunStatistics_timeoutActions - Total number of Actions that timed out.

$sel:succeededActions:WorkflowRunStatistics', workflowRunStatistics_succeededActions - Total number of Actions that have succeeded.

workflowRunStatistics_runningActions :: Lens' WorkflowRunStatistics (Maybe Int)

Total number Actions in running state.

workflowRunStatistics_stoppedActions :: Lens' WorkflowRunStatistics (Maybe Int)

Total number of Actions that have stopped.

workflowRunStatistics_totalActions :: Lens' WorkflowRunStatistics (Maybe Int)

Total number of Actions in the workflow run.

workflowRunStatistics_failedActions :: Lens' WorkflowRunStatistics (Maybe Int)

Total number of Actions that have failed.

workflowRunStatistics_timeoutActions :: Lens' WorkflowRunStatistics (Maybe Int)

Total number of Actions that timed out.

workflowRunStatistics_succeededActions :: Lens' WorkflowRunStatistics (Maybe Int)

Total number of Actions that have succeeded.