Network.AWS.Glue.UpdateColumnStatisticsForPartition

Description

Creates or updates partition statistics of columns.

The Identity and Access Management (IAM) permission required for this operation is UpdatePartition.

Creating a Request

data UpdateColumnStatisticsForPartition

See: newUpdateColumnStatisticsForPartition smart constructor.

Constructors

UpdateColumnStatisticsForPartition'

Fields

catalogId :: Maybe Text

The ID of the Data Catalog where the partitions in question reside. If none is supplied, the Amazon Web Services account ID is used by default.

databaseName :: Text

The name of the catalog database where the partitions reside.

tableName :: Text

The name of the partitions' table.

partitionValues :: [Text]

A list of partition values identifying the partition.

columnStatisticsList :: [ColumnStatistics]

A list of the column statistics.

newUpdateColumnStatisticsForPartition

Arguments

:: Text
$sel:databaseName:UpdateColumnStatisticsForPartition'

-> Text
$sel:tableName:UpdateColumnStatisticsForPartition'

-> UpdateColumnStatisticsForPartition

Create a value of UpdateColumnStatisticsForPartition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:UpdateColumnStatisticsForPartition', updateColumnStatisticsForPartition_catalogId - The ID of the Data Catalog where the partitions in question reside. If none is supplied, the Amazon Web Services account ID is used by default.

$sel:databaseName:UpdateColumnStatisticsForPartition', updateColumnStatisticsForPartition_databaseName - The name of the catalog database where the partitions reside.

$sel:tableName:UpdateColumnStatisticsForPartition', updateColumnStatisticsForPartition_tableName - The name of the partitions' table.

$sel:partitionValues:UpdateColumnStatisticsForPartition', updateColumnStatisticsForPartition_partitionValues - A list of partition values identifying the partition.

$sel:columnStatisticsList:UpdateColumnStatisticsForPartition', updateColumnStatisticsForPartition_columnStatisticsList - A list of the column statistics.

Request Lenses

updateColumnStatisticsForPartition_catalogId :: Lens' UpdateColumnStatisticsForPartition (Maybe Text)

The ID of the Data Catalog where the partitions in question reside. If none is supplied, the Amazon Web Services account ID is used by default.

updateColumnStatisticsForPartition_databaseName :: Lens' UpdateColumnStatisticsForPartition Text

The name of the catalog database where the partitions reside.

updateColumnStatisticsForPartition_tableName :: Lens' UpdateColumnStatisticsForPartition Text

The name of the partitions' table.

updateColumnStatisticsForPartition_partitionValues :: Lens' UpdateColumnStatisticsForPartition [Text]

A list of partition values identifying the partition.

updateColumnStatisticsForPartition_columnStatisticsList :: Lens' UpdateColumnStatisticsForPartition [ColumnStatistics]

A list of the column statistics.

Destructuring the Response

data UpdateColumnStatisticsForPartitionResponse

See: newUpdateColumnStatisticsForPartitionResponse smart constructor.

Constructors

UpdateColumnStatisticsForPartitionResponse'

Fields

errors :: Maybe [ColumnStatisticsError]

Error occurred during updating column statistics data.

httpStatus :: Int

The response's http status code.

newUpdateColumnStatisticsForPartitionResponse

Arguments

:: Int
$sel:httpStatus:UpdateColumnStatisticsForPartitionResponse'

-> UpdateColumnStatisticsForPartitionResponse

Create a value of UpdateColumnStatisticsForPartitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errors:UpdateColumnStatisticsForPartitionResponse', updateColumnStatisticsForPartitionResponse_errors - Error occurred during updating column statistics data.

$sel:httpStatus:UpdateColumnStatisticsForPartitionResponse', updateColumnStatisticsForPartitionResponse_httpStatus - The response's http status code.

Response Lenses

updateColumnStatisticsForPartitionResponse_errors :: Lens' UpdateColumnStatisticsForPartitionResponse (Maybe [ColumnStatisticsError])

Error occurred during updating column statistics data.

updateColumnStatisticsForPartitionResponse_httpStatus :: Lens' UpdateColumnStatisticsForPartitionResponse Int

The response's http status code.