Network.AWS.Glue.UpdateCrawler

Description

Updates a crawler. If a crawler is running, you must stop it using StopCrawler before updating it.

Creating a Request

data UpdateCrawler

See: newUpdateCrawler smart constructor.

Constructors

UpdateCrawler'

Fields

schemaChangePolicy :: Maybe SchemaChangePolicy

The policy for the crawler's update and deletion behavior.

recrawlPolicy :: Maybe RecrawlPolicy

A policy that specifies whether to crawl the entire dataset again, or to crawl only folders that were added since the last crawler run.

classifiers :: Maybe [Text]

A list of custom classifiers that the user has registered. By default, all built-in classifiers are included in a crawl, but these custom classifiers always override the default classifiers for a given classification.

configuration :: Maybe Text

Crawler configuration information. This versioned JSON string allows users to specify aspects of a crawler's behavior. For more information, see Configuring a Crawler.

lineageConfiguration :: Maybe LineageConfiguration

Specifies data lineage configuration settings for the crawler.

targets :: Maybe CrawlerTargets

A list of targets to crawl.

role' :: Maybe Text

The IAM role or Amazon Resource Name (ARN) of an IAM role that is used by the new crawler to access customer resources.

description :: Maybe Text

A description of the new crawler.

tablePrefix :: Maybe Text

The table prefix used for catalog tables that are created.

schedule :: Maybe Text

A cron expression used to specify the schedule (see Time-Based Schedules for Jobs and Crawlers. For example, to run something every day at 12:15 UTC, you would specify: cron(15 12 * * ? *).

crawlerSecurityConfiguration :: Maybe Text

The name of the SecurityConfiguration structure to be used by this crawler.

databaseName :: Maybe Text

The Glue database where results are stored, such as: arn:aws:daylight:us-east-1::database/sometable/*.

name :: Text

Name of the new crawler.

newUpdateCrawler

Arguments

:: Text
$sel:name:UpdateCrawler'

-> UpdateCrawler

Create a value of UpdateCrawler with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schemaChangePolicy:UpdateCrawler', updateCrawler_schemaChangePolicy - The policy for the crawler's update and deletion behavior.

$sel:recrawlPolicy:UpdateCrawler', updateCrawler_recrawlPolicy - A policy that specifies whether to crawl the entire dataset again, or to crawl only folders that were added since the last crawler run.

$sel:classifiers:UpdateCrawler', updateCrawler_classifiers - A list of custom classifiers that the user has registered. By default, all built-in classifiers are included in a crawl, but these custom classifiers always override the default classifiers for a given classification.

$sel:configuration:UpdateCrawler', updateCrawler_configuration - Crawler configuration information. This versioned JSON string allows users to specify aspects of a crawler's behavior. For more information, see Configuring a Crawler.

$sel:lineageConfiguration:UpdateCrawler', updateCrawler_lineageConfiguration - Specifies data lineage configuration settings for the crawler.

$sel:targets:UpdateCrawler', updateCrawler_targets - A list of targets to crawl.

$sel:role':UpdateCrawler', updateCrawler_role - The IAM role or Amazon Resource Name (ARN) of an IAM role that is used by the new crawler to access customer resources.

$sel:description:UpdateCrawler', updateCrawler_description - A description of the new crawler.

$sel:tablePrefix:UpdateCrawler', updateCrawler_tablePrefix - The table prefix used for catalog tables that are created.

$sel:schedule:UpdateCrawler', updateCrawler_schedule - A cron expression used to specify the schedule (see Time-Based Schedules for Jobs and Crawlers. For example, to run something every day at 12:15 UTC, you would specify: cron(15 12 * * ? *).

$sel:crawlerSecurityConfiguration:UpdateCrawler', updateCrawler_crawlerSecurityConfiguration - The name of the SecurityConfiguration structure to be used by this crawler.

$sel:databaseName:UpdateCrawler', updateCrawler_databaseName - The Glue database where results are stored, such as: arn:aws:daylight:us-east-1::database/sometable/*.

$sel:name:UpdateCrawler', updateCrawler_name - Name of the new crawler.

Request Lenses

updateCrawler_schemaChangePolicy :: Lens' UpdateCrawler (Maybe SchemaChangePolicy)

The policy for the crawler's update and deletion behavior.

updateCrawler_recrawlPolicy :: Lens' UpdateCrawler (Maybe RecrawlPolicy)

A policy that specifies whether to crawl the entire dataset again, or to crawl only folders that were added since the last crawler run.

updateCrawler_classifiers :: Lens' UpdateCrawler (Maybe [Text])

A list of custom classifiers that the user has registered. By default, all built-in classifiers are included in a crawl, but these custom classifiers always override the default classifiers for a given classification.

updateCrawler_configuration :: Lens' UpdateCrawler (Maybe Text)

Crawler configuration information. This versioned JSON string allows users to specify aspects of a crawler's behavior. For more information, see Configuring a Crawler.

updateCrawler_lineageConfiguration :: Lens' UpdateCrawler (Maybe LineageConfiguration)

Specifies data lineage configuration settings for the crawler.

updateCrawler_targets :: Lens' UpdateCrawler (Maybe CrawlerTargets)

A list of targets to crawl.

updateCrawler_role :: Lens' UpdateCrawler (Maybe Text)

The IAM role or Amazon Resource Name (ARN) of an IAM role that is used by the new crawler to access customer resources.

updateCrawler_description :: Lens' UpdateCrawler (Maybe Text)

A description of the new crawler.

updateCrawler_tablePrefix :: Lens' UpdateCrawler (Maybe Text)

The table prefix used for catalog tables that are created.

updateCrawler_schedule :: Lens' UpdateCrawler (Maybe Text)

A cron expression used to specify the schedule (see Time-Based Schedules for Jobs and Crawlers. For example, to run something every day at 12:15 UTC, you would specify: cron(15 12 * * ? *).

updateCrawler_crawlerSecurityConfiguration :: Lens' UpdateCrawler (Maybe Text)

The name of the SecurityConfiguration structure to be used by this crawler.

updateCrawler_databaseName :: Lens' UpdateCrawler (Maybe Text)

The Glue database where results are stored, such as: arn:aws:daylight:us-east-1::database/sometable/*.

updateCrawler_name :: Lens' UpdateCrawler Text

Name of the new crawler.

Destructuring the Response

data UpdateCrawlerResponse

See: newUpdateCrawlerResponse smart constructor.

Constructors

UpdateCrawlerResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateCrawlerResponse

Arguments

:: Int
$sel:httpStatus:UpdateCrawlerResponse'

-> UpdateCrawlerResponse

Create a value of UpdateCrawlerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateCrawlerResponse', updateCrawlerResponse_httpStatus - The response's http status code.

Response Lenses

updateCrawlerResponse_httpStatus :: Lens' UpdateCrawlerResponse Int

The response's http status code.