Network.AWS.Glue.UpdateDatabase

Description

Updates an existing database definition in a Data Catalog.

Creating a Request

data UpdateDatabase

See: newUpdateDatabase smart constructor.

Constructors

UpdateDatabase'

Fields

catalogId :: Maybe Text

The ID of the Data Catalog in which the metadata database resides. If none is provided, the Amazon Web Services account ID is used by default.

name :: Text

The name of the database to update in the catalog. For Hive compatibility, this is folded to lowercase.

databaseInput :: DatabaseInput

A DatabaseInput object specifying the new definition of the metadata database in the catalog.

newUpdateDatabase

Arguments

:: Text
$sel:name:UpdateDatabase'

-> DatabaseInput
$sel:databaseInput:UpdateDatabase'

-> UpdateDatabase

Create a value of UpdateDatabase with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:UpdateDatabase', updateDatabase_catalogId - The ID of the Data Catalog in which the metadata database resides. If none is provided, the Amazon Web Services account ID is used by default.

$sel:name:UpdateDatabase', updateDatabase_name - The name of the database to update in the catalog. For Hive compatibility, this is folded to lowercase.

$sel:databaseInput:UpdateDatabase', updateDatabase_databaseInput - A DatabaseInput object specifying the new definition of the metadata database in the catalog.

Request Lenses

updateDatabase_catalogId :: Lens' UpdateDatabase (Maybe Text)

The ID of the Data Catalog in which the metadata database resides. If none is provided, the Amazon Web Services account ID is used by default.

updateDatabase_name :: Lens' UpdateDatabase Text

The name of the database to update in the catalog. For Hive compatibility, this is folded to lowercase.

updateDatabase_databaseInput :: Lens' UpdateDatabase DatabaseInput

A DatabaseInput object specifying the new definition of the metadata database in the catalog.

Destructuring the Response

data UpdateDatabaseResponse

See: newUpdateDatabaseResponse smart constructor.

Constructors

UpdateDatabaseResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateDatabaseResponse

Arguments

:: Int
$sel:httpStatus:UpdateDatabaseResponse'

-> UpdateDatabaseResponse

Create a value of UpdateDatabaseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateDatabaseResponse', updateDatabaseResponse_httpStatus - The response's http status code.

Response Lenses

updateDatabaseResponse_httpStatus :: Lens' UpdateDatabaseResponse Int

The response's http status code.