Network.AWS.Glue.UpdateDevEndpoint

Description

Updates a specified development endpoint.

Creating a Request

data UpdateDevEndpoint

See: newUpdateDevEndpoint smart constructor.

Constructors

UpdateDevEndpoint'

Fields

publicKey :: Maybe Text

The public key for the DevEndpoint to use.

updateEtlLibraries :: Maybe Bool

True if the list of custom libraries to be loaded in the development endpoint needs to be updated, or False if otherwise.

addPublicKeys :: Maybe [Text]

The list of public keys for the DevEndpoint to use.

deletePublicKeys :: Maybe [Text]

The list of public keys to be deleted from the DevEndpoint.

deleteArguments :: Maybe [Text]

The list of argument keys to be deleted from the map of arguments used to configure the DevEndpoint.

addArguments :: Maybe (HashMap Text Text)

The map of arguments to add the map of arguments used to configure the DevEndpoint.

Valid arguments are:

"--enable-glue-datacatalog": ""

You can specify a version of Python support for development endpoints by using the Arguments parameter in the CreateDevEndpoint or UpdateDevEndpoint APIs. If no arguments are provided, the version defaults to Python 2.

customLibraries :: Maybe DevEndpointCustomLibraries

Custom Python or Java libraries to be loaded in the DevEndpoint.

endpointName :: Text

The name of the DevEndpoint to be updated.

newUpdateDevEndpoint

Arguments

:: Text
$sel:endpointName:UpdateDevEndpoint'

-> UpdateDevEndpoint

Create a value of UpdateDevEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:publicKey:UpdateDevEndpoint', updateDevEndpoint_publicKey - The public key for the DevEndpoint to use.

$sel:updateEtlLibraries:UpdateDevEndpoint', updateDevEndpoint_updateEtlLibraries - True if the list of custom libraries to be loaded in the development endpoint needs to be updated, or False if otherwise.

$sel:addPublicKeys:UpdateDevEndpoint', updateDevEndpoint_addPublicKeys - The list of public keys for the DevEndpoint to use.

$sel:deletePublicKeys:UpdateDevEndpoint', updateDevEndpoint_deletePublicKeys - The list of public keys to be deleted from the DevEndpoint.

$sel:deleteArguments:UpdateDevEndpoint', updateDevEndpoint_deleteArguments - The list of argument keys to be deleted from the map of arguments used to configure the DevEndpoint.

$sel:addArguments:UpdateDevEndpoint', updateDevEndpoint_addArguments - The map of arguments to add the map of arguments used to configure the DevEndpoint.

Valid arguments are:

"--enable-glue-datacatalog": ""

You can specify a version of Python support for development endpoints by using the Arguments parameter in the CreateDevEndpoint or UpdateDevEndpoint APIs. If no arguments are provided, the version defaults to Python 2.

$sel:customLibraries:UpdateDevEndpoint', updateDevEndpoint_customLibraries - Custom Python or Java libraries to be loaded in the DevEndpoint.

$sel:endpointName:UpdateDevEndpoint', updateDevEndpoint_endpointName - The name of the DevEndpoint to be updated.

Request Lenses

updateDevEndpoint_publicKey :: Lens' UpdateDevEndpoint (Maybe Text)

The public key for the DevEndpoint to use.

updateDevEndpoint_updateEtlLibraries :: Lens' UpdateDevEndpoint (Maybe Bool)

True if the list of custom libraries to be loaded in the development endpoint needs to be updated, or False if otherwise.

updateDevEndpoint_addPublicKeys :: Lens' UpdateDevEndpoint (Maybe [Text])

The list of public keys for the DevEndpoint to use.

updateDevEndpoint_deletePublicKeys :: Lens' UpdateDevEndpoint (Maybe [Text])

The list of public keys to be deleted from the DevEndpoint.

updateDevEndpoint_deleteArguments :: Lens' UpdateDevEndpoint (Maybe [Text])

The list of argument keys to be deleted from the map of arguments used to configure the DevEndpoint.

updateDevEndpoint_addArguments :: Lens' UpdateDevEndpoint (Maybe (HashMap Text Text))

The map of arguments to add the map of arguments used to configure the DevEndpoint.

Valid arguments are:

"--enable-glue-datacatalog": ""

You can specify a version of Python support for development endpoints by using the Arguments parameter in the CreateDevEndpoint or UpdateDevEndpoint APIs. If no arguments are provided, the version defaults to Python 2.

updateDevEndpoint_customLibraries :: Lens' UpdateDevEndpoint (Maybe DevEndpointCustomLibraries)

Custom Python or Java libraries to be loaded in the DevEndpoint.

updateDevEndpoint_endpointName :: Lens' UpdateDevEndpoint Text

The name of the DevEndpoint to be updated.

Destructuring the Response

data UpdateDevEndpointResponse

See: newUpdateDevEndpointResponse smart constructor.

Constructors

UpdateDevEndpointResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateDevEndpointResponse

Arguments

:: Int
$sel:httpStatus:UpdateDevEndpointResponse'

-> UpdateDevEndpointResponse

Create a value of UpdateDevEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateDevEndpointResponse', updateDevEndpointResponse_httpStatus - The response's http status code.

Response Lenses

updateDevEndpointResponse_httpStatus :: Lens' UpdateDevEndpointResponse Int

The response's http status code.