Network.AWS.Glue.UpdateMLTransform

Description

Updates an existing machine learning transform. Call this operation to tune the algorithm parameters to achieve better results.

After calling this operation, you can call the StartMLEvaluationTaskRun operation to assess how well your new parameters achieved your goals (such as improving the quality of your machine learning transform, or making it more cost-effective).

Creating a Request

data UpdateMLTransform

See: newUpdateMLTransform smart constructor.

Constructors

UpdateMLTransform'

Fields

numberOfWorkers :: Maybe Int

The number of workers of a defined workerType that are allocated when this task runs.

workerType :: Maybe WorkerType

The type of predefined worker that is allocated when this task runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker provides 4 vCPU, 16 GB of memory and a 64GB disk, and 1 executor per worker.

For the G.2X worker type, each worker provides 8 vCPU, 32 GB of memory and a 128GB disk, and 1 executor per worker.

glueVersion :: Maybe Text

This value determines which version of Glue this machine learning transform is compatible with. Glue 1.0 is recommended for most customers. If the value is not set, the Glue compatibility defaults to Glue 0.9. For more information, see Glue Versions in the developer guide.

role' :: Maybe Text

The name or Amazon Resource Name (ARN) of the IAM role with the required permissions.

name :: Maybe Text

The unique name that you gave the transform when you created it.

parameters :: Maybe TransformParameters

The configuration parameters that are specific to the transform type (algorithm) used. Conditionally dependent on the transform type.

maxRetries :: Maybe Int

The maximum number of times to retry a task for this transform after a task run fails.

maxCapacity :: Maybe Double

The number of Glue data processing units (DPUs) that are allocated to task runs for this transform. You can allocate from 2 to 100 DPUs; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

When the WorkerType field is set to a value other than Standard, the MaxCapacity field is set automatically and becomes read-only.

timeout :: Maybe Natural

The timeout for a task run for this transform in minutes. This is the maximum time that a task run for this transform can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours).

description :: Maybe Text

A description of the transform. The default is an empty string.

transformId :: Text

A unique identifier that was generated when the transform was created.

newUpdateMLTransform

Arguments

:: Text
$sel:transformId:UpdateMLTransform'

-> UpdateMLTransform

Create a value of UpdateMLTransform with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:numberOfWorkers:UpdateMLTransform', updateMLTransform_numberOfWorkers - The number of workers of a defined workerType that are allocated when this task runs.

$sel:workerType:UpdateMLTransform', updateMLTransform_workerType - The type of predefined worker that is allocated when this task runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker provides 4 vCPU, 16 GB of memory and a 64GB disk, and 1 executor per worker.

For the G.2X worker type, each worker provides 8 vCPU, 32 GB of memory and a 128GB disk, and 1 executor per worker.

$sel:glueVersion:UpdateMLTransform', updateMLTransform_glueVersion - This value determines which version of Glue this machine learning transform is compatible with. Glue 1.0 is recommended for most customers. If the value is not set, the Glue compatibility defaults to Glue 0.9. For more information, see Glue Versions in the developer guide.

$sel:role':UpdateMLTransform', updateMLTransform_role - The name or Amazon Resource Name (ARN) of the IAM role with the required permissions.

$sel:name:UpdateMLTransform', updateMLTransform_name - The unique name that you gave the transform when you created it.

$sel:parameters:UpdateMLTransform', updateMLTransform_parameters - The configuration parameters that are specific to the transform type (algorithm) used. Conditionally dependent on the transform type.

$sel:maxRetries:UpdateMLTransform', updateMLTransform_maxRetries - The maximum number of times to retry a task for this transform after a task run fails.

$sel:maxCapacity:UpdateMLTransform', updateMLTransform_maxCapacity - The number of Glue data processing units (DPUs) that are allocated to task runs for this transform. You can allocate from 2 to 100 DPUs; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

When the WorkerType field is set to a value other than Standard, the MaxCapacity field is set automatically and becomes read-only.

$sel:timeout:UpdateMLTransform', updateMLTransform_timeout - The timeout for a task run for this transform in minutes. This is the maximum time that a task run for this transform can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours).

$sel:description:UpdateMLTransform', updateMLTransform_description - A description of the transform. The default is an empty string.

$sel:transformId:UpdateMLTransform', updateMLTransform_transformId - A unique identifier that was generated when the transform was created.

Request Lenses

updateMLTransform_numberOfWorkers :: Lens' UpdateMLTransform (Maybe Int)

The number of workers of a defined workerType that are allocated when this task runs.

updateMLTransform_workerType :: Lens' UpdateMLTransform (Maybe WorkerType)

The type of predefined worker that is allocated when this task runs. Accepts a value of Standard, G.1X, or G.2X.

For the Standard worker type, each worker provides 4 vCPU, 16 GB of memory and a 50GB disk, and 2 executors per worker.

For the G.1X worker type, each worker provides 4 vCPU, 16 GB of memory and a 64GB disk, and 1 executor per worker.

For the G.2X worker type, each worker provides 8 vCPU, 32 GB of memory and a 128GB disk, and 1 executor per worker.

updateMLTransform_glueVersion :: Lens' UpdateMLTransform (Maybe Text)

This value determines which version of Glue this machine learning transform is compatible with. Glue 1.0 is recommended for most customers. If the value is not set, the Glue compatibility defaults to Glue 0.9. For more information, see Glue Versions in the developer guide.

updateMLTransform_role :: Lens' UpdateMLTransform (Maybe Text)

The name or Amazon Resource Name (ARN) of the IAM role with the required permissions.

updateMLTransform_name :: Lens' UpdateMLTransform (Maybe Text)

The unique name that you gave the transform when you created it.

updateMLTransform_parameters :: Lens' UpdateMLTransform (Maybe TransformParameters)

The configuration parameters that are specific to the transform type (algorithm) used. Conditionally dependent on the transform type.

updateMLTransform_maxRetries :: Lens' UpdateMLTransform (Maybe Int)

The maximum number of times to retry a task for this transform after a task run fails.

updateMLTransform_maxCapacity :: Lens' UpdateMLTransform (Maybe Double)

The number of Glue data processing units (DPUs) that are allocated to task runs for this transform. You can allocate from 2 to 100 DPUs; the default is 10. A DPU is a relative measure of processing power that consists of 4 vCPUs of compute capacity and 16 GB of memory. For more information, see the Glue pricing page.

When the WorkerType field is set to a value other than Standard, the MaxCapacity field is set automatically and becomes read-only.

updateMLTransform_timeout :: Lens' UpdateMLTransform (Maybe Natural)

The timeout for a task run for this transform in minutes. This is the maximum time that a task run for this transform can consume resources before it is terminated and enters TIMEOUT status. The default is 2,880 minutes (48 hours).

updateMLTransform_description :: Lens' UpdateMLTransform (Maybe Text)

A description of the transform. The default is an empty string.

updateMLTransform_transformId :: Lens' UpdateMLTransform Text

A unique identifier that was generated when the transform was created.

Destructuring the Response

data UpdateMLTransformResponse

See: newUpdateMLTransformResponse smart constructor.

Constructors

UpdateMLTransformResponse'

Fields

transformId :: Maybe Text

The unique identifier for the transform that was updated.

httpStatus :: Int

The response's http status code.

newUpdateMLTransformResponse

Arguments

:: Int
$sel:httpStatus:UpdateMLTransformResponse'

-> UpdateMLTransformResponse

Create a value of UpdateMLTransformResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transformId:UpdateMLTransform', updateMLTransformResponse_transformId - The unique identifier for the transform that was updated.

$sel:httpStatus:UpdateMLTransformResponse', updateMLTransformResponse_httpStatus - The response's http status code.

Response Lenses

updateMLTransformResponse_transformId :: Lens' UpdateMLTransformResponse (Maybe Text)

The unique identifier for the transform that was updated.

updateMLTransformResponse_httpStatus :: Lens' UpdateMLTransformResponse Int

The response's http status code.