Network.AWS.Glue.UpdateTable

Description

Updates a metadata table in the Data Catalog.

Creating a Request

data UpdateTable

See: newUpdateTable smart constructor.

Constructors

UpdateTable'

Fields

catalogId :: Maybe Text

The ID of the Data Catalog where the table resides. If none is provided, the Amazon Web Services account ID is used by default.

skipArchive :: Maybe Bool

By default, UpdateTable always creates an archived version of the table before updating it. However, if skipArchive is set to true, UpdateTable does not create the archived version.

databaseName :: Text

The name of the catalog database in which the table resides. For Hive compatibility, this name is entirely lowercase.

tableInput :: TableInput

An updated TableInput object to define the metadata table in the catalog.

newUpdateTable

Arguments

:: Text
$sel:databaseName:UpdateTable'

-> TableInput
$sel:tableInput:UpdateTable'

-> UpdateTable

Create a value of UpdateTable with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:UpdateTable', updateTable_catalogId - The ID of the Data Catalog where the table resides. If none is provided, the Amazon Web Services account ID is used by default.

$sel:skipArchive:UpdateTable', updateTable_skipArchive - By default, UpdateTable always creates an archived version of the table before updating it. However, if skipArchive is set to true, UpdateTable does not create the archived version.

$sel:databaseName:UpdateTable', updateTable_databaseName - The name of the catalog database in which the table resides. For Hive compatibility, this name is entirely lowercase.

$sel:tableInput:UpdateTable', updateTable_tableInput - An updated TableInput object to define the metadata table in the catalog.

Request Lenses

updateTable_catalogId :: Lens' UpdateTable (Maybe Text)

The ID of the Data Catalog where the table resides. If none is provided, the Amazon Web Services account ID is used by default.

updateTable_skipArchive :: Lens' UpdateTable (Maybe Bool)

By default, UpdateTable always creates an archived version of the table before updating it. However, if skipArchive is set to true, UpdateTable does not create the archived version.

updateTable_databaseName :: Lens' UpdateTable Text

The name of the catalog database in which the table resides. For Hive compatibility, this name is entirely lowercase.

updateTable_tableInput :: Lens' UpdateTable TableInput

An updated TableInput object to define the metadata table in the catalog.

Destructuring the Response

data UpdateTableResponse

See: newUpdateTableResponse smart constructor.

Constructors

UpdateTableResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateTableResponse

Arguments

:: Int
$sel:httpStatus:UpdateTableResponse'

-> UpdateTableResponse

Create a value of UpdateTableResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateTableResponse', updateTableResponse_httpStatus - The response's http status code.

Response Lenses

updateTableResponse_httpStatus :: Lens' UpdateTableResponse Int

The response's http status code.