Network.AWS.Grafana

Description

Derived from API version 2020-08-18 of the AWS service descriptions, licensed under Apache 2.0.

Amazon Managed Grafana is a fully managed and secure data visualization service that you can use to instantly query, correlate, and visualize operational metrics, logs, and traces from multiple sources. Amazon Managed Grafana makes it easy to deploy, operate, and scale Grafana, a widely deployed data visualization tool that is popular for its extensible data support.

With Amazon Managed Grafana, you create logically isolated Grafana servers called workspaces. In a workspace, you can create Grafana dashboards and visualizations to analyze your metrics, logs, and traces without having to build, package, or deploy any hardware to run Grafana servers.

Service Configuration

defaultService :: Service

API version 2020-08-18 of the Amazon Managed Grafana SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by Grafana.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The value of a parameter in the request caused an error.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient permissions to perform this action.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

A resource was in an inconsistent state during an update or a deletion.

ServiceQuotaExceededException

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The request would cause a service quota to be exceeded.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied because of request throttling. Retry the request.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

Unexpected error while processing the request. Retry the request.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The request references a resource that does not exist.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListWorkspaces (Paginated)

data ListWorkspaces

See: newListWorkspaces smart constructor.

Constructors

ListWorkspaces' (Maybe Text) (Maybe Natural)

newListWorkspaces :: ListWorkspaces

Create a value of ListWorkspaces with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListWorkspaces', listWorkspaces_nextToken - The token for the next set of workspaces to return. (You receive this token from a previous ListWorkspaces operation.)

$sel:maxResults:ListWorkspaces', listWorkspaces_maxResults - The maximum number of workspaces to include in the results.

data ListWorkspacesResponse

See: newListWorkspacesResponse smart constructor.

Constructors

ListWorkspacesResponse' (Maybe Text) Int [WorkspaceSummary]

newListWorkspacesResponse

Arguments

:: Int
$sel:httpStatus:ListWorkspacesResponse'

-> ListWorkspacesResponse

Create a value of ListWorkspacesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListWorkspaces', listWorkspacesResponse_nextToken - The token to use when requesting the next set of workspaces.

$sel:httpStatus:ListWorkspacesResponse', listWorkspacesResponse_httpStatus - The response's http status code.

$sel:workspaces:ListWorkspacesResponse', listWorkspacesResponse_workspaces - An array of structures that contain some information about the workspaces in the account.

DeleteWorkspace

data DeleteWorkspace

See: newDeleteWorkspace smart constructor.

Constructors

DeleteWorkspace' Text

newDeleteWorkspace

Arguments

:: Text
$sel:workspaceId:DeleteWorkspace'

-> DeleteWorkspace

Create a value of DeleteWorkspace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workspaceId:DeleteWorkspace', deleteWorkspace_workspaceId - The ID of the workspace to delete.

data DeleteWorkspaceResponse

See: newDeleteWorkspaceResponse smart constructor.

Constructors

DeleteWorkspaceResponse' Int WorkspaceDescription

newDeleteWorkspaceResponse

Arguments

:: Int
$sel:httpStatus:DeleteWorkspaceResponse'

-> WorkspaceDescription
$sel:workspace:DeleteWorkspaceResponse'

-> DeleteWorkspaceResponse

Create a value of DeleteWorkspaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteWorkspaceResponse', deleteWorkspaceResponse_httpStatus - The response's http status code.

$sel:workspace:DeleteWorkspaceResponse', deleteWorkspaceResponse_workspace - A structure containing information about the workspace that was deleted.

UpdateWorkspace

data UpdateWorkspace

See: newUpdateWorkspace smart constructor.

Constructors

UpdateWorkspace' (Maybe (Sensitive Text)) (Maybe [DataSourceType]) (Maybe PermissionType) (Maybe [NotificationDestinationType]) (Maybe (Sensitive Text)) (Maybe AccountAccessType) (Maybe (Sensitive Text)) (Maybe (Sensitive Text)) (Maybe Text) (Maybe (Sensitive [Text])) Text

newUpdateWorkspace

Arguments

:: Text
$sel:workspaceId:UpdateWorkspace'

-> UpdateWorkspace

Create a value of UpdateWorkspace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workspaceRoleArn:UpdateWorkspace', updateWorkspace_workspaceRoleArn - The workspace needs an IAM role that grants permissions to the Amazon Web Services resources that the workspace will view data from. If you already have a role that you want to use, specify it here. If you omit this field and you specify some Amazon Web Services resources in workspaceDataSources or workspaceNotificationDestinations, a new IAM role with the necessary permissions is automatically created.

$sel:workspaceDataSources:UpdateWorkspace', updateWorkspace_workspaceDataSources - Specify the Amazon Web Services data sources that you want to be queried in this workspace. Specifying these data sources here enables Amazon Managed Grafana to create IAM roles and permissions that allow Amazon Managed Grafana to read data from these sources. You must still add them as data sources in the Grafana console in the workspace.

If you don't specify a data source here, you can still add it as a data source later in the workspace console. However, you will then have to manually configure permissions for it.

$sel:permissionType:UpdateWorkspace', updateWorkspace_permissionType - If you specify Service Managed, Amazon Managed Grafana automatically creates the IAM roles and provisions the permissions that the workspace needs to use Amazon Web Services data sources and notification channels.

If you specify CUSTOMER_MANAGED, you will manage those roles and permissions yourself. If you are creating this workspace in a member account of an organization and that account is not a delegated administrator account, and you want the workspace to access data sources in other Amazon Web Services accounts in the organization, you must choose CUSTOMER_MANAGED.

For more information, see Amazon Managed Grafana permissions and policies for Amazon Web Services data sources and notification channels

$sel:workspaceNotificationDestinations:UpdateWorkspace', updateWorkspace_workspaceNotificationDestinations - Specify the Amazon Web Services notification channels that you plan to use in this workspace. Specifying these data sources here enables Amazon Managed Grafana to create IAM roles and permissions that allow Amazon Managed Grafana to use these channels.

$sel:workspaceName:UpdateWorkspace', updateWorkspace_workspaceName - A new name for the workspace to update.

$sel:accountAccessType:UpdateWorkspace', updateWorkspace_accountAccessType - Specifies whether the workspace can access Amazon Web Services resources in this Amazon Web Services account only, or whether it can also access Amazon Web Services resources in other accounts in the same organization. If you specify ORGANIZATION, you must specify which organizational units the workspace can access in the workspaceOrganizationalUnits parameter.

$sel:organizationRoleName:UpdateWorkspace', updateWorkspace_organizationRoleName - The name of an IAM role that already exists to use to access resources through Organizations.

$sel:workspaceDescription:UpdateWorkspace', updateWorkspace_workspaceDescription - A description for the workspace. This is used only to help you identify this workspace.

$sel:stackSetName:UpdateWorkspace', updateWorkspace_stackSetName - The name of the CloudFormation stack set to use to generate IAM roles to be used for this workspace.

$sel:workspaceOrganizationalUnits:UpdateWorkspace', updateWorkspace_workspaceOrganizationalUnits - Specifies the organizational units that this workspace is allowed to use data sources from, if this workspace is in an account that is part of an organization.

$sel:workspaceId:UpdateWorkspace', updateWorkspace_workspaceId - The ID of the workspace to update.

data UpdateWorkspaceResponse

See: newUpdateWorkspaceResponse smart constructor.

Constructors

UpdateWorkspaceResponse' Int WorkspaceDescription

newUpdateWorkspaceResponse

Arguments

:: Int
$sel:httpStatus:UpdateWorkspaceResponse'

-> WorkspaceDescription
$sel:workspace:UpdateWorkspaceResponse'

-> UpdateWorkspaceResponse

Create a value of UpdateWorkspaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateWorkspaceResponse', updateWorkspaceResponse_httpStatus - The response's http status code.

$sel:workspace:UpdateWorkspaceResponse', updateWorkspaceResponse_workspace - A structure containing data about the workspace that was created.

UpdateWorkspaceAuthentication

data UpdateWorkspaceAuthentication

See: newUpdateWorkspaceAuthentication smart constructor.

Constructors

UpdateWorkspaceAuthentication' (Maybe SamlConfiguration) [AuthenticationProviderTypes] Text

newUpdateWorkspaceAuthentication

Arguments

:: Text
$sel:workspaceId:UpdateWorkspaceAuthentication'

-> UpdateWorkspaceAuthentication

Create a value of UpdateWorkspaceAuthentication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:samlConfiguration:UpdateWorkspaceAuthentication', updateWorkspaceAuthentication_samlConfiguration - If the workspace uses SAML, use this structure to map SAML assertion attributes to workspace user information and define which groups in the assertion attribute are to have the Admin and Editor roles in the workspace.

$sel:authenticationProviders:UpdateWorkspaceAuthentication', updateWorkspaceAuthentication_authenticationProviders - Specifies whether this workspace uses SAML 2.0, Amazon Web Services Single Sign On, or both to authenticate users for using the Grafana console within a workspace. For more information, see User authentication in Amazon Managed Grafana.

$sel:workspaceId:UpdateWorkspaceAuthentication', updateWorkspaceAuthentication_workspaceId - The ID of the workspace to update the authentication for.

data UpdateWorkspaceAuthenticationResponse

See: newUpdateWorkspaceAuthenticationResponse smart constructor.

Constructors

UpdateWorkspaceAuthenticationResponse' Int AuthenticationDescription

newUpdateWorkspaceAuthenticationResponse

Arguments

:: Int
$sel:httpStatus:UpdateWorkspaceAuthenticationResponse'

-> AuthenticationDescription
$sel:authentication:UpdateWorkspaceAuthenticationResponse'

-> UpdateWorkspaceAuthenticationResponse

Create a value of UpdateWorkspaceAuthenticationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateWorkspaceAuthenticationResponse', updateWorkspaceAuthenticationResponse_httpStatus - The response's http status code.

$sel:authentication:UpdateWorkspaceAuthenticationResponse', updateWorkspaceAuthenticationResponse_authentication - A structure that describes the user authentication for this workspace after the update is made.

DescribeWorkspaceAuthentication

data DescribeWorkspaceAuthentication

See: newDescribeWorkspaceAuthentication smart constructor.

Constructors

DescribeWorkspaceAuthentication' Text

newDescribeWorkspaceAuthentication

Arguments

:: Text
$sel:workspaceId:DescribeWorkspaceAuthentication'

-> DescribeWorkspaceAuthentication

Create a value of DescribeWorkspaceAuthentication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workspaceId:DescribeWorkspaceAuthentication', describeWorkspaceAuthentication_workspaceId - The ID of the workspace to return authentication information about.

data DescribeWorkspaceAuthenticationResponse

See: newDescribeWorkspaceAuthenticationResponse smart constructor.

Constructors

DescribeWorkspaceAuthenticationResponse' Int AuthenticationDescription

newDescribeWorkspaceAuthenticationResponse

Arguments

:: Int
$sel:httpStatus:DescribeWorkspaceAuthenticationResponse'

-> AuthenticationDescription
$sel:authentication:DescribeWorkspaceAuthenticationResponse'

-> DescribeWorkspaceAuthenticationResponse

Create a value of DescribeWorkspaceAuthenticationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeWorkspaceAuthenticationResponse', describeWorkspaceAuthenticationResponse_httpStatus - The response's http status code.

$sel:authentication:DescribeWorkspaceAuthenticationResponse', describeWorkspaceAuthenticationResponse_authentication - A structure containing information about the authentication methods used in the workspace.

DescribeWorkspace

data DescribeWorkspace

See: newDescribeWorkspace smart constructor.

Constructors

DescribeWorkspace' Text

newDescribeWorkspace

Arguments

:: Text
$sel:workspaceId:DescribeWorkspace'

-> DescribeWorkspace

Create a value of DescribeWorkspace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workspaceId:DescribeWorkspace', describeWorkspace_workspaceId - The ID of the workspace to display information about.

data DescribeWorkspaceResponse

See: newDescribeWorkspaceResponse smart constructor.

Constructors

DescribeWorkspaceResponse' Int WorkspaceDescription

newDescribeWorkspaceResponse

Arguments

:: Int
$sel:httpStatus:DescribeWorkspaceResponse'

-> WorkspaceDescription
$sel:workspace:DescribeWorkspaceResponse'

-> DescribeWorkspaceResponse

Create a value of DescribeWorkspaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeWorkspaceResponse', describeWorkspaceResponse_httpStatus - The response's http status code.

$sel:workspace:DescribeWorkspaceResponse', describeWorkspaceResponse_workspace - A structure containing information about the workspace.

AssociateLicense

data AssociateLicense

See: newAssociateLicense smart constructor.

Constructors

AssociateLicense' LicenseType Text

newAssociateLicense

Arguments

:: LicenseType
$sel:licenseType:AssociateLicense'

-> Text
$sel:workspaceId:AssociateLicense'

-> AssociateLicense

Create a value of AssociateLicense with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:licenseType:AssociateLicense', associateLicense_licenseType - The type of license to associate with the workspace.

$sel:workspaceId:AssociateLicense', associateLicense_workspaceId - The ID of the workspace to associate the license with.

data AssociateLicenseResponse

See: newAssociateLicenseResponse smart constructor.

Constructors

AssociateLicenseResponse' Int WorkspaceDescription

newAssociateLicenseResponse

Arguments

:: Int
$sel:httpStatus:AssociateLicenseResponse'

-> WorkspaceDescription
$sel:workspace:AssociateLicenseResponse'

-> AssociateLicenseResponse

Create a value of AssociateLicenseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AssociateLicenseResponse', associateLicenseResponse_httpStatus - The response's http status code.

$sel:workspace:AssociateLicenseResponse', associateLicenseResponse_workspace - A structure containing data about the workspace.

ListPermissions (Paginated)

data ListPermissions

See: newListPermissions smart constructor.

Constructors

ListPermissions' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Natural) (Maybe UserType) Text

newListPermissions

Arguments

:: Text
$sel:workspaceId:ListPermissions'

-> ListPermissions

Create a value of ListPermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userId:ListPermissions', listPermissions_userId - (Optional) Limits the results to only the user that matches this ID.

$sel:nextToken:ListPermissions', listPermissions_nextToken - The token to use when requesting the next set of results. You received this token from a previous ListPermissions operation.

$sel:groupId:ListPermissions', listPermissions_groupId - (Optional) Limits the results to only the group that matches this ID.

$sel:maxResults:ListPermissions', listPermissions_maxResults - The maximum number of results to include in the response.

$sel:userType:ListPermissions', listPermissions_userType - (Optional) If you specify SSO_USER, then only the permissions of Amazon Web Services SSO users are returned. If you specify SSO_GROUP, only the permissions of Amazon Web Services SSO groups are returned.

$sel:workspaceId:ListPermissions', listPermissions_workspaceId - The ID of the workspace to list permissions for. This parameter is required.

data ListPermissionsResponse

See: newListPermissionsResponse smart constructor.

Constructors

ListPermissionsResponse' (Maybe Text) Int [PermissionEntry]

newListPermissionsResponse

Arguments

:: Int
$sel:httpStatus:ListPermissionsResponse'

-> ListPermissionsResponse

Create a value of ListPermissionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPermissions', listPermissionsResponse_nextToken - The token to use in a subsequent ListPermissions operation to return the next set of results.

$sel:httpStatus:ListPermissionsResponse', listPermissionsResponse_httpStatus - The response's http status code.

$sel:permissions:ListPermissionsResponse', listPermissionsResponse_permissions - The permissions returned by the operation.

UpdatePermissions

data UpdatePermissions

See: newUpdatePermissions smart constructor.

Constructors

UpdatePermissions' [UpdateInstruction] Text

newUpdatePermissions

Arguments

:: Text
$sel:workspaceId:UpdatePermissions'

-> UpdatePermissions

Create a value of UpdatePermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateInstructionBatch:UpdatePermissions', updatePermissions_updateInstructionBatch - An array of structures that contain the permission updates to make.

$sel:workspaceId:UpdatePermissions', updatePermissions_workspaceId - The ID of the workspace to update.

data UpdatePermissionsResponse

See: newUpdatePermissionsResponse smart constructor.

Constructors

UpdatePermissionsResponse' Int [UpdateError]

newUpdatePermissionsResponse

Arguments

:: Int
$sel:httpStatus:UpdatePermissionsResponse'

-> UpdatePermissionsResponse

Create a value of UpdatePermissionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdatePermissionsResponse', updatePermissionsResponse_httpStatus - The response's http status code.

$sel:errors:UpdatePermissionsResponse', updatePermissionsResponse_errors - An array of structures that contain the errors from the operation, if any.

DisassociateLicense

data DisassociateLicense

See: newDisassociateLicense smart constructor.

Constructors

DisassociateLicense' LicenseType Text

newDisassociateLicense

Arguments

:: LicenseType
$sel:licenseType:DisassociateLicense'

-> Text
$sel:workspaceId:DisassociateLicense'

-> DisassociateLicense

Create a value of DisassociateLicense with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:licenseType:DisassociateLicense', disassociateLicense_licenseType - The type of license to remove from the workspace.

$sel:workspaceId:DisassociateLicense', disassociateLicense_workspaceId - The ID of the workspace to remove the Grafana Enterprise license from.

data DisassociateLicenseResponse

See: newDisassociateLicenseResponse smart constructor.

Constructors

DisassociateLicenseResponse' Int WorkspaceDescription

newDisassociateLicenseResponse

Arguments

:: Int
$sel:httpStatus:DisassociateLicenseResponse'

-> WorkspaceDescription
$sel:workspace:DisassociateLicenseResponse'

-> DisassociateLicenseResponse

Create a value of DisassociateLicenseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisassociateLicenseResponse', disassociateLicenseResponse_httpStatus - The response's http status code.

$sel:workspace:DisassociateLicenseResponse', disassociateLicenseResponse_workspace - A structure containing information about the workspace.

CreateWorkspace

data CreateWorkspace

See: newCreateWorkspace smart constructor.

Constructors

CreateWorkspace' (Maybe (Sensitive Text)) (Maybe Text) (Maybe [DataSourceType]) (Maybe [NotificationDestinationType]) (Maybe (Sensitive Text)) (Maybe (Sensitive Text)) (Maybe (Sensitive Text)) (Maybe Text) (Maybe (Sensitive [Text])) AccountAccessType [AuthenticationProviderTypes] PermissionType

newCreateWorkspace

Arguments

:: AccountAccessType
$sel:accountAccessType:CreateWorkspace'

-> PermissionType
$sel:permissionType:CreateWorkspace'

-> CreateWorkspace

Create a value of CreateWorkspace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workspaceRoleArn:CreateWorkspace', createWorkspace_workspaceRoleArn - The workspace needs an IAM role that grants permissions to the Amazon Web Services resources that the workspace will view data from. If you already have a role that you want to use, specify it here. If you omit this field and you specify some Amazon Web Services resources in workspaceDataSources or workspaceNotificationDestinations, a new IAM role with the necessary permissions is automatically created.

$sel:clientToken:CreateWorkspace', createWorkspace_clientToken - A unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

$sel:workspaceDataSources:CreateWorkspace', createWorkspace_workspaceDataSources - Specify the Amazon Web Services data sources that you want to be queried in this workspace. Specifying these data sources here enables Amazon Managed Grafana to create IAM roles and permissions that allow Amazon Managed Grafana to read data from these sources. You must still add them as data sources in the Grafana console in the workspace.

If you don't specify a data source here, you can still add it as a data source in the workspace console later. However, you will then have to manually configure permissions for it.

$sel:workspaceNotificationDestinations:CreateWorkspace', createWorkspace_workspaceNotificationDestinations - Specify the Amazon Web Services notification channels that you plan to use in this workspace. Specifying these data sources here enables Amazon Managed Grafana to create IAM roles and permissions that allow Amazon Managed Grafana to use these channels.

$sel:workspaceName:CreateWorkspace', createWorkspace_workspaceName - The name for the workspace. It does not have to be unique.

$sel:organizationRoleName:CreateWorkspace', createWorkspace_organizationRoleName - The name of an IAM role that already exists to use with Organizations to access Amazon Web Services data sources and notification channels in other accounts in an organization.

$sel:workspaceDescription:CreateWorkspace', createWorkspace_workspaceDescription - A description for the workspace. This is used only to help you identify this workspace.

$sel:stackSetName:CreateWorkspace', createWorkspace_stackSetName - The name of the CloudFormation stack set to use to generate IAM roles to be used for this workspace.

$sel:workspaceOrganizationalUnits:CreateWorkspace', createWorkspace_workspaceOrganizationalUnits - Specifies the organizational units that this workspace is allowed to use data sources from, if this workspace is in an account that is part of an organization.

$sel:accountAccessType:CreateWorkspace', createWorkspace_accountAccessType - Specifies whether the workspace can access Amazon Web Services resources in this Amazon Web Services account only, or whether it can also access Amazon Web Services resources in other accounts in the same organization. If you specify ORGANIZATION, you must specify which organizational units the workspace can access in the workspaceOrganizationalUnits parameter.

$sel:authenticationProviders:CreateWorkspace', createWorkspace_authenticationProviders - Specifies whether this workspace uses SAML 2.0, Amazon Web Services Single Sign On, or both to authenticate users for using the Grafana console within a workspace. For more information, see User authentication in Amazon Managed Grafana.

$sel:permissionType:CreateWorkspace', createWorkspace_permissionType - If you specify Service Managed, Amazon Managed Grafana automatically creates the IAM roles and provisions the permissions that the workspace needs to use Amazon Web Services data sources and notification channels.

If you specify CUSTOMER_MANAGED, you will manage those roles and permissions yourself. If you are creating this workspace in a member account of an organization that is not a delegated administrator account, and you want the workspace to access data sources in other Amazon Web Services accounts in the organization, you must choose CUSTOMER_MANAGED.

For more information, see Amazon Managed Grafana permissions and policies for Amazon Web Services data sources and notification channels

data CreateWorkspaceResponse

See: newCreateWorkspaceResponse smart constructor.

Constructors

CreateWorkspaceResponse' Int WorkspaceDescription

newCreateWorkspaceResponse

Arguments

:: Int
$sel:httpStatus:CreateWorkspaceResponse'

-> WorkspaceDescription
$sel:workspace:CreateWorkspaceResponse'

-> CreateWorkspaceResponse

Create a value of CreateWorkspaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateWorkspaceResponse', createWorkspaceResponse_httpStatus - The response's http status code.

$sel:workspace:CreateWorkspaceResponse', createWorkspaceResponse_workspace - A structure containing data about the workspace that was created.

Types

AccountAccessType

newtype AccountAccessType

Constructors

AccountAccessType'

Fields

fromAccountAccessType :: Text

Bundled Patterns

pattern AccountAccessType_CURRENT_ACCOUNT :: AccountAccessType

pattern AccountAccessType_ORGANIZATION :: AccountAccessType

AuthenticationProviderTypes

newtype AuthenticationProviderTypes

Constructors

AuthenticationProviderTypes'

Fields

fromAuthenticationProviderTypes :: Text

Bundled Patterns

pattern AuthenticationProviderTypes_AWS_SSO :: AuthenticationProviderTypes

pattern AuthenticationProviderTypes_SAML :: AuthenticationProviderTypes

DataSourceType

newtype DataSourceType

Constructors

DataSourceType'

Fields

fromDataSourceType :: Text

Bundled Patterns

pattern DataSourceType_AMAZON_OPENSEARCH_SERVICE :: DataSourceType

pattern DataSourceType_CLOUDWATCH :: DataSourceType

pattern DataSourceType_PROMETHEUS :: DataSourceType

pattern DataSourceType_SITEWISE :: DataSourceType

pattern DataSourceType_TIMESTREAM :: DataSourceType

pattern DataSourceType_XRAY :: DataSourceType

LicenseType

newtype LicenseType

Constructors

LicenseType'

Fields

fromLicenseType :: Text

Bundled Patterns

pattern LicenseType_ENTERPRISE :: LicenseType

pattern LicenseType_ENTERPRISE_FREE_TRIAL :: LicenseType

NotificationDestinationType

newtype NotificationDestinationType

Constructors

NotificationDestinationType'

Fields

fromNotificationDestinationType :: Text

Bundled Patterns

pattern NotificationDestinationType_SNS :: NotificationDestinationType

PermissionType

newtype PermissionType

Constructors

PermissionType'

Fields

fromPermissionType :: Text

Bundled Patterns

pattern PermissionType_CUSTOMER_MANAGED :: PermissionType

pattern PermissionType_SERVICE_MANAGED :: PermissionType

Role

newtype Role

Constructors

Role'

Fields

fromRole :: Text

Bundled Patterns

pattern Role_ADMIN :: Role

pattern Role_EDITOR :: Role

SamlConfigurationStatus

newtype SamlConfigurationStatus

Constructors

SamlConfigurationStatus'

Fields

fromSamlConfigurationStatus :: Text

Bundled Patterns

pattern SamlConfigurationStatus_CONFIGURED :: SamlConfigurationStatus

pattern SamlConfigurationStatus_NOT_CONFIGURED :: SamlConfigurationStatus

UpdateAction

newtype UpdateAction

Constructors

UpdateAction'

Fields

fromUpdateAction :: Text

Bundled Patterns

pattern UpdateAction_ADD :: UpdateAction

pattern UpdateAction_REVOKE :: UpdateAction

UserType

newtype UserType

Constructors

UserType'

Fields

fromUserType :: Text

Bundled Patterns

pattern UserType_SSO_GROUP :: UserType

pattern UserType_SSO_USER :: UserType

WorkspaceStatus

newtype WorkspaceStatus

Constructors

WorkspaceStatus'

Fields

fromWorkspaceStatus :: Text

Bundled Patterns

pattern WorkspaceStatus_ACTIVE :: WorkspaceStatus

pattern WorkspaceStatus_CREATING :: WorkspaceStatus

pattern WorkspaceStatus_CREATION_FAILED :: WorkspaceStatus

pattern WorkspaceStatus_DELETING :: WorkspaceStatus

pattern WorkspaceStatus_DELETION_FAILED :: WorkspaceStatus

pattern WorkspaceStatus_FAILED :: WorkspaceStatus

pattern WorkspaceStatus_LICENSE_REMOVAL_FAILED :: WorkspaceStatus

pattern WorkspaceStatus_UPDATE_FAILED :: WorkspaceStatus

pattern WorkspaceStatus_UPDATING :: WorkspaceStatus

pattern WorkspaceStatus_UPGRADE_FAILED :: WorkspaceStatus

pattern WorkspaceStatus_UPGRADING :: WorkspaceStatus

AssertionAttributes

data AssertionAttributes

A structure that defines which attributes in the IdP assertion are to be used to define information about the users authenticated by the IdP to use the workspace.

See: newAssertionAttributes smart constructor.

Constructors

AssertionAttributes' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text)

newAssertionAttributes :: AssertionAttributes

Create a value of AssertionAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:email:AssertionAttributes', assertionAttributes_email - The name of the attribute within the SAML assertion to use as the email names for SAML users.

$sel:groups:AssertionAttributes', assertionAttributes_groups - The name of the attribute within the SAML assertion to use as the user full "friendly" names for user groups.

$sel:org:AssertionAttributes', assertionAttributes_org - The name of the attribute within the SAML assertion to use as the user full "friendly" names for the users' organizations.

$sel:role':AssertionAttributes', assertionAttributes_role - The name of the attribute within the SAML assertion to use as the user roles.

$sel:name:AssertionAttributes', assertionAttributes_name - The name of the attribute within the SAML assertion to use as the user full "friendly" names for SAML users.

$sel:login:AssertionAttributes', assertionAttributes_login - The name of the attribute within the SAML assertion to use as the login names for SAML users.

AuthenticationDescription

data AuthenticationDescription

A structure containing information about the user authentication methods used by the workspace.

See: newAuthenticationDescription smart constructor.

Constructors

AuthenticationDescription' (Maybe AwsSsoAuthentication) (Maybe SamlAuthentication) [AuthenticationProviderTypes]

newAuthenticationDescription :: AuthenticationDescription

Create a value of AuthenticationDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsSso:AuthenticationDescription', authenticationDescription_awsSso - A structure containing information about how this workspace works with Amazon Web Services SSO.

$sel:saml:AuthenticationDescription', authenticationDescription_saml - A structure containing information about how this workspace works with SAML, including what attributes within the assertion are to be mapped to user information in the workspace.

$sel:providers:AuthenticationDescription', authenticationDescription_providers - Specifies whether this workspace uses Amazon Web Services SSO, SAML, or both methods to authenticate users to use the Grafana console in the Amazon Managed Grafana workspace.

AuthenticationSummary

data AuthenticationSummary

A structure that describes whether the workspace uses SAML, Amazon Web Services SSO, or both methods for user authentication, and whether that authentication is fully configured.

See: newAuthenticationSummary smart constructor.

Constructors

AuthenticationSummary' (Maybe SamlConfigurationStatus) [AuthenticationProviderTypes]

newAuthenticationSummary :: AuthenticationSummary

Create a value of AuthenticationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:samlConfigurationStatus:AuthenticationSummary', authenticationSummary_samlConfigurationStatus - Specifies whether the workplace's user authentication method is fully configured.

$sel:providers:AuthenticationSummary', authenticationSummary_providers - Specifies whether the workspace uses SAML, Amazon Web Services SSO, or both methods for user authentication.

AwsSsoAuthentication

data AwsSsoAuthentication

A structure containing information about how this workspace works with Amazon Web Services SSO.

See: newAwsSsoAuthentication smart constructor.

Constructors

AwsSsoAuthentication' (Maybe Text)

newAwsSsoAuthentication :: AwsSsoAuthentication

Create a value of AwsSsoAuthentication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ssoClientId:AwsSsoAuthentication', awsSsoAuthentication_ssoClientId - The ID of the Amazon Web Services SSO-managed application that is created by Amazon Managed Grafana.

IdpMetadata

data IdpMetadata

A structure containing the identity provider (IdP) metadata used to integrate the identity provider with this workspace. You can specify the metadata either by providing a URL to its location in the url parameter, or by specifying the full metadata in XML format in the xml parameter.

See: newIdpMetadata smart constructor.

Constructors

IdpMetadata' (Maybe Text) (Maybe Text)

newIdpMetadata :: IdpMetadata

Create a value of IdpMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:url:IdpMetadata', idpMetadata_url - The URL of the location containing the metadata.

$sel:xml:IdpMetadata', idpMetadata_xml - The actual full metadata file, in XML format.

PermissionEntry

data PermissionEntry

A structure containing the identity of one user or group and the Admin or Editor role that they have.

See: newPermissionEntry smart constructor.

Constructors

PermissionEntry' Role User

newPermissionEntry

Arguments

:: Role
$sel:role':PermissionEntry'

-> User
$sel:user:PermissionEntry'

-> PermissionEntry

Create a value of PermissionEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:role':PermissionEntry', permissionEntry_role - Specifies whether the user or group has the Admin or Editor role.

$sel:user:PermissionEntry', permissionEntry_user - A structure with the ID of the user or group with this role.

RoleValues

data RoleValues

This structure defines which groups defined in the SAML assertion attribute are to be mapped to the Grafana Admin and Editor roles in the workspace.

See: newRoleValues smart constructor.

Constructors

RoleValues' (Maybe [Text]) (Maybe [Text])

newRoleValues :: RoleValues

Create a value of RoleValues with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:admin:RoleValues', roleValues_admin - A list of groups from the SAML assertion attribute to grant the Grafana Admin role to.

$sel:editor:RoleValues', roleValues_editor - A list of groups from the SAML assertion attribute to grant the Grafana Editor role to.

SamlAuthentication

data SamlAuthentication

A structure containing information about how this workspace works with SAML.

See: newSamlAuthentication smart constructor.

Constructors

SamlAuthentication' (Maybe SamlConfiguration) SamlConfigurationStatus

newSamlAuthentication

Arguments

:: SamlConfigurationStatus
$sel:status:SamlAuthentication'

-> SamlAuthentication

Create a value of SamlAuthentication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configuration:SamlAuthentication', samlAuthentication_configuration - A structure containing details about how this workspace works with SAML.

$sel:status:SamlAuthentication', samlAuthentication_status - Specifies whether the workspace's SAML configuration is complete.

SamlConfiguration

data SamlConfiguration

A structure containing information about how this workspace works with SAML.

See: newSamlConfiguration smart constructor.

Constructors

SamlConfiguration' (Maybe Int) (Maybe AssertionAttributes) (Maybe [Text]) (Maybe RoleValues) IdpMetadata

newSamlConfiguration

Arguments

:: IdpMetadata
$sel:idpMetadata:SamlConfiguration'

-> SamlConfiguration

Create a value of SamlConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loginValidityDuration:SamlConfiguration', samlConfiguration_loginValidityDuration - How long a sign-on session by a SAML user is valid, before the user has to sign on again.

$sel:assertionAttributes:SamlConfiguration', samlConfiguration_assertionAttributes - A structure that defines which attributes in the SAML assertion are to be used to define information about the users authenticated by that IdP to use the workspace.

$sel:allowedOrganizations:SamlConfiguration', samlConfiguration_allowedOrganizations - Lists which organizations defined in the SAML assertion are allowed to use the Amazon Managed Grafana workspace. If this is empty, all organizations in the assertion attribute have access.

$sel:roleValues:SamlConfiguration', samlConfiguration_roleValues - A structure containing arrays that map group names in the SAML assertion to the Grafana Admin and Editor roles in the workspace.

$sel:idpMetadata:SamlConfiguration', samlConfiguration_idpMetadata - A structure containing the identity provider (IdP) metadata used to integrate the identity provider with this workspace.

UpdateError

data UpdateError

A structure containing information about one error encountered while performing an UpdatePermissions operation.

See: newUpdateError smart constructor.

Constructors

UpdateError' UpdateInstruction Natural Text

newUpdateError

Arguments

:: UpdateInstruction
$sel:causedBy:UpdateError'

-> Natural
$sel:code:UpdateError'

-> Text
$sel:message:UpdateError'

-> UpdateError

Create a value of UpdateError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:causedBy:UpdateError', updateError_causedBy - Specifies which permission update caused the error.

$sel:code:UpdateError', updateError_code - The error code.

$sel:message:UpdateError', updateError_message - The message for this error.

UpdateInstruction

data UpdateInstruction

Contains the instructions for one Grafana role permission update in a UpdatePermissions operation.

See: newUpdateInstruction smart constructor.

Constructors

UpdateInstruction' UpdateAction Role [User]

newUpdateInstruction

Arguments

:: UpdateAction
$sel:action:UpdateInstruction'

-> Role
$sel:role':UpdateInstruction'

-> UpdateInstruction

Create a value of UpdateInstruction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:action:UpdateInstruction', updateInstruction_action - Specifies whether this update is to add or revoke role permissions.

$sel:role':UpdateInstruction', updateInstruction_role - The role to add or revoke for the user or the group specified in users.

$sel:users:UpdateInstruction', updateInstruction_users - A structure that specifies the user or group to add or revoke the role for.

User

data User

A structure that specifies one user or group in the workspace.

See: newUser smart constructor.

Constructors

User' Text UserType

newUser

Arguments

:: Text
$sel:id:User'

-> UserType
$sel:type':User'

-> User

Create a value of User with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:User', user_id - The ID of the user or group.

$sel:type':User', user_type - Specifies whether this is a single user or a group.

WorkspaceDescription

data WorkspaceDescription

A structure containing information about an Amazon Managed Grafana workspace in your account.

See: newWorkspaceDescription smart constructor.

Constructors

WorkspaceDescription' (Maybe (Sensitive Text)) (Maybe POSIX) (Maybe LicenseType) (Maybe PermissionType) (Maybe (Sensitive Text)) (Maybe [NotificationDestinationType]) (Maybe AccountAccessType) (Maybe POSIX) (Maybe (Sensitive Text)) (Maybe Text) (Maybe (Sensitive [Text])) (Maybe (Sensitive Text)) (Maybe Bool) AuthenticationSummary POSIX [DataSourceType] Text Text Text POSIX WorkspaceStatus

newWorkspaceDescription

Arguments

:: AuthenticationSummary
$sel:authentication:WorkspaceDescription'

-> UTCTime
$sel:created:WorkspaceDescription'

-> Text
$sel:endpoint:WorkspaceDescription'

-> Text
$sel:grafanaVersion:WorkspaceDescription'

-> Text
$sel:id:WorkspaceDescription'

-> UTCTime
$sel:modified:WorkspaceDescription'

-> WorkspaceStatus
$sel:status:WorkspaceDescription'

-> WorkspaceDescription

Create a value of WorkspaceDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workspaceRoleArn:WorkspaceDescription', workspaceDescription_workspaceRoleArn - The IAM role that grants permissions to the Amazon Web Services resources that the workspace will view data from. This role must already exist.

$sel:freeTrialExpiration:WorkspaceDescription', workspaceDescription_freeTrialExpiration - If this workspace is currently in the free trial period for Grafana Enterprise, this value specifies when that free trial ends.

$sel:licenseType:WorkspaceDescription', workspaceDescription_licenseType - Specifies whether this workspace has a full Grafana Enterprise license or a free trial license.

$sel:permissionType:WorkspaceDescription', workspaceDescription_permissionType - If this is Service Managed, Amazon Managed Grafana automatically creates the IAM roles and provisions the permissions that the workspace needs to use Amazon Web Services data sources and notification channels.

If this is CUSTOMER_MANAGED, you manage those roles and permissions yourself. If you are creating this workspace in a member account of an organization and that account is not a delegated administrator account, and you want the workspace to access data sources in other Amazon Web Services accounts in the organization, you must choose CUSTOMER_MANAGED.

For more information, see Amazon Managed Grafana permissions and policies for Amazon Web Services data sources and notification channels

$sel:name:WorkspaceDescription', workspaceDescription_name - The name of the workspace.

$sel:notificationDestinations:WorkspaceDescription', workspaceDescription_notificationDestinations - The Amazon Web Services notification channels that Amazon Managed Grafana can automatically create IAM roles and permissions for, to allow Amazon Managed Grafana to use these channels.

$sel:accountAccessType:WorkspaceDescription', workspaceDescription_accountAccessType - Specifies whether the workspace can access Amazon Web Services resources in this Amazon Web Services account only, or whether it can also access Amazon Web Services resources in other accounts in the same organization. If this is ORGANIZATION, the workspaceOrganizationalUnits parameter specifies which organizational units the workspace can access.

$sel:licenseExpiration:WorkspaceDescription', workspaceDescription_licenseExpiration - If this workspace has a full Grafana Enterprise license, this specifies when the license ends and will need to be renewed.

$sel:organizationRoleName:WorkspaceDescription', workspaceDescription_organizationRoleName - The name of the IAM role that is used to access resources through Organizations.

$sel:stackSetName:WorkspaceDescription', workspaceDescription_stackSetName - The name of the CloudFormation stack set that is used to generate IAM roles to be used for this workspace.

$sel:organizationalUnits:WorkspaceDescription', workspaceDescription_organizationalUnits - Specifies the organizational units that this workspace is allowed to use data sources from, if this workspace is in an account that is part of an organization.

$sel:description:WorkspaceDescription', workspaceDescription_description - The user-defined description of the workspace.

$sel:freeTrialConsumed:WorkspaceDescription', workspaceDescription_freeTrialConsumed - Specifies whether this workspace has already fully used its free trial for Grafana Enterprise.

$sel:authentication:WorkspaceDescription', workspaceDescription_authentication - A structure that describes whether the workspace uses SAML, Amazon Web Services SSO, or both methods for user authentication.

$sel:created:WorkspaceDescription', workspaceDescription_created - The date that the workspace was created.

$sel:dataSources:WorkspaceDescription', workspaceDescription_dataSources - Specifies the Amazon Web Services data sources that have been configured to have IAM roles and permissions created to allow Amazon Managed Grafana to read data from these sources.

$sel:endpoint:WorkspaceDescription', workspaceDescription_endpoint - The URL that users can use to access the Grafana console in the workspace.

$sel:grafanaVersion:WorkspaceDescription', workspaceDescription_grafanaVersion - The version of Grafana supported in this workspace.

$sel:id:WorkspaceDescription', workspaceDescription_id - The unique ID of this workspace.

$sel:modified:WorkspaceDescription', workspaceDescription_modified - The most recent date that the workspace was modified.

$sel:status:WorkspaceDescription', workspaceDescription_status - The current status of the workspace.

WorkspaceSummary

data WorkspaceSummary

A structure that contains some information about one workspace in the account.

See: newWorkspaceSummary smart constructor.

Constructors

WorkspaceSummary' (Maybe (Sensitive Text)) (Maybe [NotificationDestinationType]) (Maybe (Sensitive Text)) AuthenticationSummary POSIX Text Text Text POSIX WorkspaceStatus

newWorkspaceSummary

Arguments

:: AuthenticationSummary
$sel:authentication:WorkspaceSummary'

-> UTCTime
$sel:created:WorkspaceSummary'

-> Text
$sel:endpoint:WorkspaceSummary'

-> Text
$sel:grafanaVersion:WorkspaceSummary'

-> Text
$sel:id:WorkspaceSummary'

-> UTCTime
$sel:modified:WorkspaceSummary'

-> WorkspaceStatus
$sel:status:WorkspaceSummary'

-> WorkspaceSummary

Create a value of WorkspaceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:WorkspaceSummary', workspaceSummary_name - The name of the workspace.

$sel:notificationDestinations:WorkspaceSummary', workspaceSummary_notificationDestinations - The Amazon Web Services notification channels that Amazon Managed Grafana can automatically create IAM roles and permissions for, which allows Amazon Managed Grafana to use these channels.

$sel:description:WorkspaceSummary', workspaceSummary_description - The customer-entered description of the workspace.

$sel:authentication:WorkspaceSummary', workspaceSummary_authentication - A structure containing information about the authentication methods used in the workspace.

$sel:created:WorkspaceSummary', workspaceSummary_created - The date that the workspace was created.

$sel:endpoint:WorkspaceSummary', workspaceSummary_endpoint - The URL endpoint to use to access the Grafana console in the workspace.

$sel:grafanaVersion:WorkspaceSummary', workspaceSummary_grafanaVersion - The Grafana version that the workspace is running.

$sel:id:WorkspaceSummary', workspaceSummary_id - The unique ID of the workspace.

$sel:modified:WorkspaceSummary', workspaceSummary_modified - The most recent date that the workspace was modified.

$sel:status:WorkspaceSummary', workspaceSummary_status - The current status of the workspace.