Network.AWS.Grafana.DisassociateLicense

Description

Removes the Grafana Enterprise license from a workspace.

Creating a Request

data DisassociateLicense

See: newDisassociateLicense smart constructor.

Constructors

DisassociateLicense'

Fields

licenseType :: LicenseType

The type of license to remove from the workspace.

workspaceId :: Text

The ID of the workspace to remove the Grafana Enterprise license from.

newDisassociateLicense

Arguments

:: LicenseType
$sel:licenseType:DisassociateLicense'

-> Text
$sel:workspaceId:DisassociateLicense'

-> DisassociateLicense

Create a value of DisassociateLicense with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:licenseType:DisassociateLicense', disassociateLicense_licenseType - The type of license to remove from the workspace.

$sel:workspaceId:DisassociateLicense', disassociateLicense_workspaceId - The ID of the workspace to remove the Grafana Enterprise license from.

Request Lenses

disassociateLicense_licenseType :: Lens' DisassociateLicense LicenseType

The type of license to remove from the workspace.

disassociateLicense_workspaceId :: Lens' DisassociateLicense Text

The ID of the workspace to remove the Grafana Enterprise license from.

Destructuring the Response

data DisassociateLicenseResponse

See: newDisassociateLicenseResponse smart constructor.

Constructors

DisassociateLicenseResponse'

Fields

httpStatus :: Int

The response's http status code.

workspace :: WorkspaceDescription

A structure containing information about the workspace.

newDisassociateLicenseResponse

Arguments

:: Int
$sel:httpStatus:DisassociateLicenseResponse'

-> WorkspaceDescription
$sel:workspace:DisassociateLicenseResponse'

-> DisassociateLicenseResponse

Create a value of DisassociateLicenseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisassociateLicenseResponse', disassociateLicenseResponse_httpStatus - The response's http status code.

$sel:workspace:DisassociateLicenseResponse', disassociateLicenseResponse_workspace - A structure containing information about the workspace.

Response Lenses

disassociateLicenseResponse_httpStatus :: Lens' DisassociateLicenseResponse Int

The response's http status code.

disassociateLicenseResponse_workspace :: Lens' DisassociateLicenseResponse WorkspaceDescription

A structure containing information about the workspace.