Network.AWS.Grafana.Lens

Description

Operations

ListWorkspaces

listWorkspaces_nextToken :: Lens' ListWorkspaces (Maybe Text)

The token for the next set of workspaces to return. (You receive this token from a previous ListWorkspaces operation.)

listWorkspaces_maxResults :: Lens' ListWorkspaces (Maybe Natural)

The maximum number of workspaces to include in the results.

listWorkspacesResponse_nextToken :: Lens' ListWorkspacesResponse (Maybe Text)

The token to use when requesting the next set of workspaces.

listWorkspacesResponse_httpStatus :: Lens' ListWorkspacesResponse Int

The response's http status code.

listWorkspacesResponse_workspaces :: Lens' ListWorkspacesResponse [WorkspaceSummary]

An array of structures that contain some information about the workspaces in the account.

DeleteWorkspace

deleteWorkspace_workspaceId :: Lens' DeleteWorkspace Text

The ID of the workspace to delete.

deleteWorkspaceResponse_httpStatus :: Lens' DeleteWorkspaceResponse Int

The response's http status code.

deleteWorkspaceResponse_workspace :: Lens' DeleteWorkspaceResponse WorkspaceDescription

A structure containing information about the workspace that was deleted.

UpdateWorkspace

updateWorkspace_workspaceRoleArn :: Lens' UpdateWorkspace (Maybe Text)

The workspace needs an IAM role that grants permissions to the Amazon Web Services resources that the workspace will view data from. If you already have a role that you want to use, specify it here. If you omit this field and you specify some Amazon Web Services resources in workspaceDataSources or workspaceNotificationDestinations, a new IAM role with the necessary permissions is automatically created.

updateWorkspace_workspaceDataSources :: Lens' UpdateWorkspace (Maybe [DataSourceType])

Specify the Amazon Web Services data sources that you want to be queried in this workspace. Specifying these data sources here enables Amazon Managed Grafana to create IAM roles and permissions that allow Amazon Managed Grafana to read data from these sources. You must still add them as data sources in the Grafana console in the workspace.

If you don't specify a data source here, you can still add it as a data source later in the workspace console. However, you will then have to manually configure permissions for it.

updateWorkspace_permissionType :: Lens' UpdateWorkspace (Maybe PermissionType)

If you specify Service Managed, Amazon Managed Grafana automatically creates the IAM roles and provisions the permissions that the workspace needs to use Amazon Web Services data sources and notification channels.

If you specify CUSTOMER_MANAGED, you will manage those roles and permissions yourself. If you are creating this workspace in a member account of an organization and that account is not a delegated administrator account, and you want the workspace to access data sources in other Amazon Web Services accounts in the organization, you must choose CUSTOMER_MANAGED.

For more information, see Amazon Managed Grafana permissions and policies for Amazon Web Services data sources and notification channels

updateWorkspace_workspaceNotificationDestinations :: Lens' UpdateWorkspace (Maybe [NotificationDestinationType])

Specify the Amazon Web Services notification channels that you plan to use in this workspace. Specifying these data sources here enables Amazon Managed Grafana to create IAM roles and permissions that allow Amazon Managed Grafana to use these channels.

updateWorkspace_workspaceName :: Lens' UpdateWorkspace (Maybe Text)

A new name for the workspace to update.

updateWorkspace_accountAccessType :: Lens' UpdateWorkspace (Maybe AccountAccessType)

Specifies whether the workspace can access Amazon Web Services resources in this Amazon Web Services account only, or whether it can also access Amazon Web Services resources in other accounts in the same organization. If you specify ORGANIZATION, you must specify which organizational units the workspace can access in the workspaceOrganizationalUnits parameter.

updateWorkspace_organizationRoleName :: Lens' UpdateWorkspace (Maybe Text)

The name of an IAM role that already exists to use to access resources through Organizations.

updateWorkspace_workspaceDescription :: Lens' UpdateWorkspace (Maybe Text)

A description for the workspace. This is used only to help you identify this workspace.

updateWorkspace_stackSetName :: Lens' UpdateWorkspace (Maybe Text)

The name of the CloudFormation stack set to use to generate IAM roles to be used for this workspace.

updateWorkspace_workspaceOrganizationalUnits :: Lens' UpdateWorkspace (Maybe [Text])

Specifies the organizational units that this workspace is allowed to use data sources from, if this workspace is in an account that is part of an organization.

updateWorkspace_workspaceId :: Lens' UpdateWorkspace Text

The ID of the workspace to update.

updateWorkspaceResponse_httpStatus :: Lens' UpdateWorkspaceResponse Int

The response's http status code.

updateWorkspaceResponse_workspace :: Lens' UpdateWorkspaceResponse WorkspaceDescription

A structure containing data about the workspace that was created.

UpdateWorkspaceAuthentication

updateWorkspaceAuthentication_samlConfiguration :: Lens' UpdateWorkspaceAuthentication (Maybe SamlConfiguration)

If the workspace uses SAML, use this structure to map SAML assertion attributes to workspace user information and define which groups in the assertion attribute are to have the Admin and Editor roles in the workspace.

updateWorkspaceAuthentication_authenticationProviders :: Lens' UpdateWorkspaceAuthentication [AuthenticationProviderTypes]

Specifies whether this workspace uses SAML 2.0, Amazon Web Services Single Sign On, or both to authenticate users for using the Grafana console within a workspace. For more information, see User authentication in Amazon Managed Grafana.

updateWorkspaceAuthentication_workspaceId :: Lens' UpdateWorkspaceAuthentication Text

The ID of the workspace to update the authentication for.

updateWorkspaceAuthenticationResponse_httpStatus :: Lens' UpdateWorkspaceAuthenticationResponse Int

The response's http status code.

updateWorkspaceAuthenticationResponse_authentication :: Lens' UpdateWorkspaceAuthenticationResponse AuthenticationDescription

A structure that describes the user authentication for this workspace after the update is made.

DescribeWorkspaceAuthentication

describeWorkspaceAuthentication_workspaceId :: Lens' DescribeWorkspaceAuthentication Text

The ID of the workspace to return authentication information about.

describeWorkspaceAuthenticationResponse_httpStatus :: Lens' DescribeWorkspaceAuthenticationResponse Int

The response's http status code.

describeWorkspaceAuthenticationResponse_authentication :: Lens' DescribeWorkspaceAuthenticationResponse AuthenticationDescription

A structure containing information about the authentication methods used in the workspace.

DescribeWorkspace

describeWorkspace_workspaceId :: Lens' DescribeWorkspace Text

The ID of the workspace to display information about.

describeWorkspaceResponse_httpStatus :: Lens' DescribeWorkspaceResponse Int

The response's http status code.

describeWorkspaceResponse_workspace :: Lens' DescribeWorkspaceResponse WorkspaceDescription

A structure containing information about the workspace.

AssociateLicense

associateLicense_licenseType :: Lens' AssociateLicense LicenseType

The type of license to associate with the workspace.

associateLicense_workspaceId :: Lens' AssociateLicense Text

The ID of the workspace to associate the license with.

associateLicenseResponse_httpStatus :: Lens' AssociateLicenseResponse Int

The response's http status code.

associateLicenseResponse_workspace :: Lens' AssociateLicenseResponse WorkspaceDescription

A structure containing data about the workspace.

ListPermissions

listPermissions_userId :: Lens' ListPermissions (Maybe Text)

(Optional) Limits the results to only the user that matches this ID.

listPermissions_nextToken :: Lens' ListPermissions (Maybe Text)

The token to use when requesting the next set of results. You received this token from a previous ListPermissions operation.

listPermissions_groupId :: Lens' ListPermissions (Maybe Text)

(Optional) Limits the results to only the group that matches this ID.

listPermissions_maxResults :: Lens' ListPermissions (Maybe Natural)

The maximum number of results to include in the response.

listPermissions_userType :: Lens' ListPermissions (Maybe UserType)

(Optional) If you specify SSO_USER, then only the permissions of Amazon Web Services SSO users are returned. If you specify SSO_GROUP, only the permissions of Amazon Web Services SSO groups are returned.

listPermissions_workspaceId :: Lens' ListPermissions Text

The ID of the workspace to list permissions for. This parameter is required.

listPermissionsResponse_nextToken :: Lens' ListPermissionsResponse (Maybe Text)

The token to use in a subsequent ListPermissions operation to return the next set of results.

listPermissionsResponse_httpStatus :: Lens' ListPermissionsResponse Int

The response's http status code.

listPermissionsResponse_permissions :: Lens' ListPermissionsResponse [PermissionEntry]

The permissions returned by the operation.

UpdatePermissions

updatePermissions_updateInstructionBatch :: Lens' UpdatePermissions [UpdateInstruction]

An array of structures that contain the permission updates to make.

updatePermissions_workspaceId :: Lens' UpdatePermissions Text

The ID of the workspace to update.

updatePermissionsResponse_httpStatus :: Lens' UpdatePermissionsResponse Int

The response's http status code.

updatePermissionsResponse_errors :: Lens' UpdatePermissionsResponse [UpdateError]

An array of structures that contain the errors from the operation, if any.

DisassociateLicense

disassociateLicense_licenseType :: Lens' DisassociateLicense LicenseType

The type of license to remove from the workspace.

disassociateLicense_workspaceId :: Lens' DisassociateLicense Text

The ID of the workspace to remove the Grafana Enterprise license from.

disassociateLicenseResponse_httpStatus :: Lens' DisassociateLicenseResponse Int

The response's http status code.

disassociateLicenseResponse_workspace :: Lens' DisassociateLicenseResponse WorkspaceDescription

A structure containing information about the workspace.

CreateWorkspace

createWorkspace_workspaceRoleArn :: Lens' CreateWorkspace (Maybe Text)

The workspace needs an IAM role that grants permissions to the Amazon Web Services resources that the workspace will view data from. If you already have a role that you want to use, specify it here. If you omit this field and you specify some Amazon Web Services resources in workspaceDataSources or workspaceNotificationDestinations, a new IAM role with the necessary permissions is automatically created.

createWorkspace_clientToken :: Lens' CreateWorkspace (Maybe Text)

A unique, case-sensitive, user-provided identifier to ensure the idempotency of the request.

createWorkspace_workspaceDataSources :: Lens' CreateWorkspace (Maybe [DataSourceType])

Specify the Amazon Web Services data sources that you want to be queried in this workspace. Specifying these data sources here enables Amazon Managed Grafana to create IAM roles and permissions that allow Amazon Managed Grafana to read data from these sources. You must still add them as data sources in the Grafana console in the workspace.

If you don't specify a data source here, you can still add it as a data source in the workspace console later. However, you will then have to manually configure permissions for it.

createWorkspace_workspaceNotificationDestinations :: Lens' CreateWorkspace (Maybe [NotificationDestinationType])

Specify the Amazon Web Services notification channels that you plan to use in this workspace. Specifying these data sources here enables Amazon Managed Grafana to create IAM roles and permissions that allow Amazon Managed Grafana to use these channels.

createWorkspace_workspaceName :: Lens' CreateWorkspace (Maybe Text)

The name for the workspace. It does not have to be unique.

createWorkspace_organizationRoleName :: Lens' CreateWorkspace (Maybe Text)

The name of an IAM role that already exists to use with Organizations to access Amazon Web Services data sources and notification channels in other accounts in an organization.

createWorkspace_workspaceDescription :: Lens' CreateWorkspace (Maybe Text)

A description for the workspace. This is used only to help you identify this workspace.

createWorkspace_stackSetName :: Lens' CreateWorkspace (Maybe Text)

The name of the CloudFormation stack set to use to generate IAM roles to be used for this workspace.

createWorkspace_workspaceOrganizationalUnits :: Lens' CreateWorkspace (Maybe [Text])

Specifies the organizational units that this workspace is allowed to use data sources from, if this workspace is in an account that is part of an organization.

createWorkspace_accountAccessType :: Lens' CreateWorkspace AccountAccessType

Specifies whether the workspace can access Amazon Web Services resources in this Amazon Web Services account only, or whether it can also access Amazon Web Services resources in other accounts in the same organization. If you specify ORGANIZATION, you must specify which organizational units the workspace can access in the workspaceOrganizationalUnits parameter.

createWorkspace_authenticationProviders :: Lens' CreateWorkspace [AuthenticationProviderTypes]

Specifies whether this workspace uses SAML 2.0, Amazon Web Services Single Sign On, or both to authenticate users for using the Grafana console within a workspace. For more information, see User authentication in Amazon Managed Grafana.

createWorkspace_permissionType :: Lens' CreateWorkspace PermissionType

If you specify Service Managed, Amazon Managed Grafana automatically creates the IAM roles and provisions the permissions that the workspace needs to use Amazon Web Services data sources and notification channels.

If you specify CUSTOMER_MANAGED, you will manage those roles and permissions yourself. If you are creating this workspace in a member account of an organization that is not a delegated administrator account, and you want the workspace to access data sources in other Amazon Web Services accounts in the organization, you must choose CUSTOMER_MANAGED.

For more information, see Amazon Managed Grafana permissions and policies for Amazon Web Services data sources and notification channels

createWorkspaceResponse_httpStatus :: Lens' CreateWorkspaceResponse Int

The response's http status code.

createWorkspaceResponse_workspace :: Lens' CreateWorkspaceResponse WorkspaceDescription

A structure containing data about the workspace that was created.

Types

AssertionAttributes

assertionAttributes_email :: Lens' AssertionAttributes (Maybe Text)

The name of the attribute within the SAML assertion to use as the email names for SAML users.

assertionAttributes_groups :: Lens' AssertionAttributes (Maybe Text)

The name of the attribute within the SAML assertion to use as the user full "friendly" names for user groups.

assertionAttributes_org :: Lens' AssertionAttributes (Maybe Text)

The name of the attribute within the SAML assertion to use as the user full "friendly" names for the users' organizations.

assertionAttributes_role :: Lens' AssertionAttributes (Maybe Text)

The name of the attribute within the SAML assertion to use as the user roles.

assertionAttributes_name :: Lens' AssertionAttributes (Maybe Text)

The name of the attribute within the SAML assertion to use as the user full "friendly" names for SAML users.

assertionAttributes_login :: Lens' AssertionAttributes (Maybe Text)

The name of the attribute within the SAML assertion to use as the login names for SAML users.

AuthenticationDescription

authenticationDescription_awsSso :: Lens' AuthenticationDescription (Maybe AwsSsoAuthentication)

A structure containing information about how this workspace works with Amazon Web Services SSO.

authenticationDescription_saml :: Lens' AuthenticationDescription (Maybe SamlAuthentication)

A structure containing information about how this workspace works with SAML, including what attributes within the assertion are to be mapped to user information in the workspace.

authenticationDescription_providers :: Lens' AuthenticationDescription [AuthenticationProviderTypes]

Specifies whether this workspace uses Amazon Web Services SSO, SAML, or both methods to authenticate users to use the Grafana console in the Amazon Managed Grafana workspace.

AuthenticationSummary

authenticationSummary_samlConfigurationStatus :: Lens' AuthenticationSummary (Maybe SamlConfigurationStatus)

Specifies whether the workplace's user authentication method is fully configured.

authenticationSummary_providers :: Lens' AuthenticationSummary [AuthenticationProviderTypes]

Specifies whether the workspace uses SAML, Amazon Web Services SSO, or both methods for user authentication.

AwsSsoAuthentication

awsSsoAuthentication_ssoClientId :: Lens' AwsSsoAuthentication (Maybe Text)

The ID of the Amazon Web Services SSO-managed application that is created by Amazon Managed Grafana.

IdpMetadata

idpMetadata_url :: Lens' IdpMetadata (Maybe Text)

The URL of the location containing the metadata.

idpMetadata_xml :: Lens' IdpMetadata (Maybe Text)

The actual full metadata file, in XML format.

PermissionEntry

permissionEntry_role :: Lens' PermissionEntry Role

Specifies whether the user or group has the Admin or Editor role.

permissionEntry_user :: Lens' PermissionEntry User

A structure with the ID of the user or group with this role.

RoleValues

roleValues_admin :: Lens' RoleValues (Maybe [Text])

A list of groups from the SAML assertion attribute to grant the Grafana Admin role to.

roleValues_editor :: Lens' RoleValues (Maybe [Text])

A list of groups from the SAML assertion attribute to grant the Grafana Editor role to.

SamlAuthentication

samlAuthentication_configuration :: Lens' SamlAuthentication (Maybe SamlConfiguration)

A structure containing details about how this workspace works with SAML.

samlAuthentication_status :: Lens' SamlAuthentication SamlConfigurationStatus

Specifies whether the workspace's SAML configuration is complete.

SamlConfiguration

samlConfiguration_loginValidityDuration :: Lens' SamlConfiguration (Maybe Int)

How long a sign-on session by a SAML user is valid, before the user has to sign on again.

samlConfiguration_assertionAttributes :: Lens' SamlConfiguration (Maybe AssertionAttributes)

A structure that defines which attributes in the SAML assertion are to be used to define information about the users authenticated by that IdP to use the workspace.

samlConfiguration_allowedOrganizations :: Lens' SamlConfiguration (Maybe [Text])

Lists which organizations defined in the SAML assertion are allowed to use the Amazon Managed Grafana workspace. If this is empty, all organizations in the assertion attribute have access.

samlConfiguration_roleValues :: Lens' SamlConfiguration (Maybe RoleValues)

A structure containing arrays that map group names in the SAML assertion to the Grafana Admin and Editor roles in the workspace.

samlConfiguration_idpMetadata :: Lens' SamlConfiguration IdpMetadata

A structure containing the identity provider (IdP) metadata used to integrate the identity provider with this workspace.

UpdateError

updateError_causedBy :: Lens' UpdateError UpdateInstruction

Specifies which permission update caused the error.

updateError_code :: Lens' UpdateError Natural

The error code.

updateError_message :: Lens' UpdateError Text

The message for this error.

UpdateInstruction

updateInstruction_action :: Lens' UpdateInstruction UpdateAction

Specifies whether this update is to add or revoke role permissions.

updateInstruction_role :: Lens' UpdateInstruction Role

The role to add or revoke for the user or the group specified in users.

updateInstruction_users :: Lens' UpdateInstruction [User]

A structure that specifies the user or group to add or revoke the role for.

User

user_id :: Lens' User Text

The ID of the user or group.

user_type :: Lens' User UserType

Specifies whether this is a single user or a group.

WorkspaceDescription

workspaceDescription_workspaceRoleArn :: Lens' WorkspaceDescription (Maybe Text)

The IAM role that grants permissions to the Amazon Web Services resources that the workspace will view data from. This role must already exist.

workspaceDescription_freeTrialExpiration :: Lens' WorkspaceDescription (Maybe UTCTime)

If this workspace is currently in the free trial period for Grafana Enterprise, this value specifies when that free trial ends.

workspaceDescription_licenseType :: Lens' WorkspaceDescription (Maybe LicenseType)

Specifies whether this workspace has a full Grafana Enterprise license or a free trial license.

workspaceDescription_permissionType :: Lens' WorkspaceDescription (Maybe PermissionType)

If this is Service Managed, Amazon Managed Grafana automatically creates the IAM roles and provisions the permissions that the workspace needs to use Amazon Web Services data sources and notification channels.

If this is CUSTOMER_MANAGED, you manage those roles and permissions yourself. If you are creating this workspace in a member account of an organization and that account is not a delegated administrator account, and you want the workspace to access data sources in other Amazon Web Services accounts in the organization, you must choose CUSTOMER_MANAGED.

For more information, see Amazon Managed Grafana permissions and policies for Amazon Web Services data sources and notification channels

workspaceDescription_name :: Lens' WorkspaceDescription (Maybe Text)

The name of the workspace.

workspaceDescription_notificationDestinations :: Lens' WorkspaceDescription (Maybe [NotificationDestinationType])

The Amazon Web Services notification channels that Amazon Managed Grafana can automatically create IAM roles and permissions for, to allow Amazon Managed Grafana to use these channels.

workspaceDescription_accountAccessType :: Lens' WorkspaceDescription (Maybe AccountAccessType)

Specifies whether the workspace can access Amazon Web Services resources in this Amazon Web Services account only, or whether it can also access Amazon Web Services resources in other accounts in the same organization. If this is ORGANIZATION, the workspaceOrganizationalUnits parameter specifies which organizational units the workspace can access.

workspaceDescription_licenseExpiration :: Lens' WorkspaceDescription (Maybe UTCTime)

If this workspace has a full Grafana Enterprise license, this specifies when the license ends and will need to be renewed.

workspaceDescription_organizationRoleName :: Lens' WorkspaceDescription (Maybe Text)

The name of the IAM role that is used to access resources through Organizations.

workspaceDescription_stackSetName :: Lens' WorkspaceDescription (Maybe Text)

The name of the CloudFormation stack set that is used to generate IAM roles to be used for this workspace.

workspaceDescription_organizationalUnits :: Lens' WorkspaceDescription (Maybe [Text])

Specifies the organizational units that this workspace is allowed to use data sources from, if this workspace is in an account that is part of an organization.

workspaceDescription_description :: Lens' WorkspaceDescription (Maybe Text)

The user-defined description of the workspace.

workspaceDescription_freeTrialConsumed :: Lens' WorkspaceDescription (Maybe Bool)

Specifies whether this workspace has already fully used its free trial for Grafana Enterprise.

workspaceDescription_authentication :: Lens' WorkspaceDescription AuthenticationSummary

A structure that describes whether the workspace uses SAML, Amazon Web Services SSO, or both methods for user authentication.

workspaceDescription_created :: Lens' WorkspaceDescription UTCTime

The date that the workspace was created.

workspaceDescription_dataSources :: Lens' WorkspaceDescription [DataSourceType]

Specifies the Amazon Web Services data sources that have been configured to have IAM roles and permissions created to allow Amazon Managed Grafana to read data from these sources.

workspaceDescription_endpoint :: Lens' WorkspaceDescription Text

The URL that users can use to access the Grafana console in the workspace.

workspaceDescription_grafanaVersion :: Lens' WorkspaceDescription Text

The version of Grafana supported in this workspace.

workspaceDescription_id :: Lens' WorkspaceDescription Text

The unique ID of this workspace.

workspaceDescription_modified :: Lens' WorkspaceDescription UTCTime

The most recent date that the workspace was modified.

workspaceDescription_status :: Lens' WorkspaceDescription WorkspaceStatus

The current status of the workspace.

WorkspaceSummary

workspaceSummary_name :: Lens' WorkspaceSummary (Maybe Text)

The name of the workspace.

workspaceSummary_notificationDestinations :: Lens' WorkspaceSummary (Maybe [NotificationDestinationType])

The Amazon Web Services notification channels that Amazon Managed Grafana can automatically create IAM roles and permissions for, which allows Amazon Managed Grafana to use these channels.

workspaceSummary_description :: Lens' WorkspaceSummary (Maybe Text)

The customer-entered description of the workspace.

workspaceSummary_authentication :: Lens' WorkspaceSummary AuthenticationSummary

A structure containing information about the authentication methods used in the workspace.

workspaceSummary_created :: Lens' WorkspaceSummary UTCTime

The date that the workspace was created.

workspaceSummary_endpoint :: Lens' WorkspaceSummary Text

The URL endpoint to use to access the Grafana console in the workspace.

workspaceSummary_grafanaVersion :: Lens' WorkspaceSummary Text

The Grafana version that the workspace is running.

workspaceSummary_id :: Lens' WorkspaceSummary Text

The unique ID of the workspace.

workspaceSummary_modified :: Lens' WorkspaceSummary UTCTime

The most recent date that the workspace was modified.

workspaceSummary_status :: Lens' WorkspaceSummary WorkspaceStatus

The current status of the workspace.