Network.AWS.Grafana.ListPermissions

Description

Lists the users and groups who have the Grafana Admin and Editor roles in this workspace. If you use this operation without specifying userId or groupId, the operation returns the roles of all users and groups. If you specify a userId or a groupId, only the roles for that user or group are returned. If you do this, you can specify only one userId or one groupId.

This operation returns paginated results.

Creating a Request

data ListPermissions

See: newListPermissions smart constructor.

Constructors

ListPermissions'

Fields

userId :: Maybe Text

(Optional) Limits the results to only the user that matches this ID.

nextToken :: Maybe Text

The token to use when requesting the next set of results. You received this token from a previous ListPermissions operation.

groupId :: Maybe Text

(Optional) Limits the results to only the group that matches this ID.

maxResults :: Maybe Natural

The maximum number of results to include in the response.

userType :: Maybe UserType

(Optional) If you specify SSO_USER, then only the permissions of Amazon Web Services SSO users are returned. If you specify SSO_GROUP, only the permissions of Amazon Web Services SSO groups are returned.

workspaceId :: Text

The ID of the workspace to list permissions for. This parameter is required.

newListPermissions

Arguments

:: Text
$sel:workspaceId:ListPermissions'

-> ListPermissions

Create a value of ListPermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userId:ListPermissions', listPermissions_userId - (Optional) Limits the results to only the user that matches this ID.

$sel:nextToken:ListPermissions', listPermissions_nextToken - The token to use when requesting the next set of results. You received this token from a previous ListPermissions operation.

$sel:groupId:ListPermissions', listPermissions_groupId - (Optional) Limits the results to only the group that matches this ID.

$sel:maxResults:ListPermissions', listPermissions_maxResults - The maximum number of results to include in the response.

$sel:userType:ListPermissions', listPermissions_userType - (Optional) If you specify SSO_USER, then only the permissions of Amazon Web Services SSO users are returned. If you specify SSO_GROUP, only the permissions of Amazon Web Services SSO groups are returned.

$sel:workspaceId:ListPermissions', listPermissions_workspaceId - The ID of the workspace to list permissions for. This parameter is required.

Request Lenses

listPermissions_userId :: Lens' ListPermissions (Maybe Text)

(Optional) Limits the results to only the user that matches this ID.

listPermissions_nextToken :: Lens' ListPermissions (Maybe Text)

The token to use when requesting the next set of results. You received this token from a previous ListPermissions operation.

listPermissions_groupId :: Lens' ListPermissions (Maybe Text)

(Optional) Limits the results to only the group that matches this ID.

listPermissions_maxResults :: Lens' ListPermissions (Maybe Natural)

The maximum number of results to include in the response.

listPermissions_userType :: Lens' ListPermissions (Maybe UserType)

(Optional) If you specify SSO_USER, then only the permissions of Amazon Web Services SSO users are returned. If you specify SSO_GROUP, only the permissions of Amazon Web Services SSO groups are returned.

listPermissions_workspaceId :: Lens' ListPermissions Text

The ID of the workspace to list permissions for. This parameter is required.

Destructuring the Response

data ListPermissionsResponse

See: newListPermissionsResponse smart constructor.

Constructors

ListPermissionsResponse'

Fields

nextToken :: Maybe Text

The token to use in a subsequent ListPermissions operation to return the next set of results.

httpStatus :: Int

The response's http status code.

permissions :: [PermissionEntry]

The permissions returned by the operation.

newListPermissionsResponse

Arguments

:: Int
$sel:httpStatus:ListPermissionsResponse'

-> ListPermissionsResponse

Create a value of ListPermissionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPermissions', listPermissionsResponse_nextToken - The token to use in a subsequent ListPermissions operation to return the next set of results.

$sel:httpStatus:ListPermissionsResponse', listPermissionsResponse_httpStatus - The response's http status code.

$sel:permissions:ListPermissionsResponse', listPermissionsResponse_permissions - The permissions returned by the operation.

Response Lenses

listPermissionsResponse_nextToken :: Lens' ListPermissionsResponse (Maybe Text)

The token to use in a subsequent ListPermissions operation to return the next set of results.

listPermissionsResponse_httpStatus :: Lens' ListPermissionsResponse Int

The response's http status code.

listPermissionsResponse_permissions :: Lens' ListPermissionsResponse [PermissionEntry]

The permissions returned by the operation.