Network.AWS.Grafana.Types.AuthenticationSummary

Description

Documentation

data AuthenticationSummary

A structure that describes whether the workspace uses SAML, Amazon Web Services SSO, or both methods for user authentication, and whether that authentication is fully configured.

See: newAuthenticationSummary smart constructor.

Constructors

AuthenticationSummary'

Fields

samlConfigurationStatus :: Maybe SamlConfigurationStatus

Specifies whether the workplace's user authentication method is fully configured.

providers :: [AuthenticationProviderTypes]

Specifies whether the workspace uses SAML, Amazon Web Services SSO, or both methods for user authentication.

newAuthenticationSummary :: AuthenticationSummary

Create a value of AuthenticationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:samlConfigurationStatus:AuthenticationSummary', authenticationSummary_samlConfigurationStatus - Specifies whether the workplace's user authentication method is fully configured.

$sel:providers:AuthenticationSummary', authenticationSummary_providers - Specifies whether the workspace uses SAML, Amazon Web Services SSO, or both methods for user authentication.

authenticationSummary_samlConfigurationStatus :: Lens' AuthenticationSummary (Maybe SamlConfigurationStatus)

Specifies whether the workplace's user authentication method is fully configured.

authenticationSummary_providers :: Lens' AuthenticationSummary [AuthenticationProviderTypes]

Specifies whether the workspace uses SAML, Amazon Web Services SSO, or both methods for user authentication.