Network.AWS.Grafana.Types.RoleValues

Description

Documentation

data RoleValues

This structure defines which groups defined in the SAML assertion attribute are to be mapped to the Grafana Admin and Editor roles in the workspace.

See: newRoleValues smart constructor.

Constructors

RoleValues'

Fields

admin :: Maybe [Text]

A list of groups from the SAML assertion attribute to grant the Grafana Admin role to.

editor :: Maybe [Text]

A list of groups from the SAML assertion attribute to grant the Grafana Editor role to.

newRoleValues :: RoleValues

Create a value of RoleValues with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:admin:RoleValues', roleValues_admin - A list of groups from the SAML assertion attribute to grant the Grafana Admin role to.

$sel:editor:RoleValues', roleValues_editor - A list of groups from the SAML assertion attribute to grant the Grafana Editor role to.

roleValues_admin :: Lens' RoleValues (Maybe [Text])

A list of groups from the SAML assertion attribute to grant the Grafana Admin role to.

roleValues_editor :: Lens' RoleValues (Maybe [Text])

A list of groups from the SAML assertion attribute to grant the Grafana Editor role to.