Network.AWS.Grafana.Types.SamlConfiguration

Description

Documentation

data SamlConfiguration

A structure containing information about how this workspace works with SAML.

See: newSamlConfiguration smart constructor.

Constructors

SamlConfiguration'

Fields

loginValidityDuration :: Maybe Int

How long a sign-on session by a SAML user is valid, before the user has to sign on again.

assertionAttributes :: Maybe AssertionAttributes

A structure that defines which attributes in the SAML assertion are to be used to define information about the users authenticated by that IdP to use the workspace.

allowedOrganizations :: Maybe [Text]

Lists which organizations defined in the SAML assertion are allowed to use the Amazon Managed Grafana workspace. If this is empty, all organizations in the assertion attribute have access.

roleValues :: Maybe RoleValues

A structure containing arrays that map group names in the SAML assertion to the Grafana Admin and Editor roles in the workspace.

idpMetadata :: IdpMetadata

A structure containing the identity provider (IdP) metadata used to integrate the identity provider with this workspace.

newSamlConfiguration

Arguments

:: IdpMetadata
$sel:idpMetadata:SamlConfiguration'

-> SamlConfiguration

Create a value of SamlConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loginValidityDuration:SamlConfiguration', samlConfiguration_loginValidityDuration - How long a sign-on session by a SAML user is valid, before the user has to sign on again.

$sel:assertionAttributes:SamlConfiguration', samlConfiguration_assertionAttributes - A structure that defines which attributes in the SAML assertion are to be used to define information about the users authenticated by that IdP to use the workspace.

$sel:allowedOrganizations:SamlConfiguration', samlConfiguration_allowedOrganizations - Lists which organizations defined in the SAML assertion are allowed to use the Amazon Managed Grafana workspace. If this is empty, all organizations in the assertion attribute have access.

$sel:roleValues:SamlConfiguration', samlConfiguration_roleValues - A structure containing arrays that map group names in the SAML assertion to the Grafana Admin and Editor roles in the workspace.

$sel:idpMetadata:SamlConfiguration', samlConfiguration_idpMetadata - A structure containing the identity provider (IdP) metadata used to integrate the identity provider with this workspace.

samlConfiguration_loginValidityDuration :: Lens' SamlConfiguration (Maybe Int)

How long a sign-on session by a SAML user is valid, before the user has to sign on again.

samlConfiguration_assertionAttributes :: Lens' SamlConfiguration (Maybe AssertionAttributes)

A structure that defines which attributes in the SAML assertion are to be used to define information about the users authenticated by that IdP to use the workspace.

samlConfiguration_allowedOrganizations :: Lens' SamlConfiguration (Maybe [Text])

Lists which organizations defined in the SAML assertion are allowed to use the Amazon Managed Grafana workspace. If this is empty, all organizations in the assertion attribute have access.

samlConfiguration_roleValues :: Lens' SamlConfiguration (Maybe RoleValues)

A structure containing arrays that map group names in the SAML assertion to the Grafana Admin and Editor roles in the workspace.

samlConfiguration_idpMetadata :: Lens' SamlConfiguration IdpMetadata

A structure containing the identity provider (IdP) metadata used to integrate the identity provider with this workspace.