Network.AWS.Grafana.Types.UpdateInstruction

Description

Documentation

data UpdateInstruction

Contains the instructions for one Grafana role permission update in a UpdatePermissions operation.

See: newUpdateInstruction smart constructor.

Constructors

UpdateInstruction'

Fields

action :: UpdateAction

Specifies whether this update is to add or revoke role permissions.

role' :: Role

The role to add or revoke for the user or the group specified in users.

users :: [User]

A structure that specifies the user or group to add or revoke the role for.

newUpdateInstruction

Arguments

:: UpdateAction
$sel:action:UpdateInstruction'

-> Role
$sel:role':UpdateInstruction'

-> UpdateInstruction

Create a value of UpdateInstruction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:action:UpdateInstruction', updateInstruction_action - Specifies whether this update is to add or revoke role permissions.

$sel:role':UpdateInstruction', updateInstruction_role - The role to add or revoke for the user or the group specified in users.

$sel:users:UpdateInstruction', updateInstruction_users - A structure that specifies the user or group to add or revoke the role for.

updateInstruction_action :: Lens' UpdateInstruction UpdateAction

Specifies whether this update is to add or revoke role permissions.

updateInstruction_role :: Lens' UpdateInstruction Role

The role to add or revoke for the user or the group specified in users.

updateInstruction_users :: Lens' UpdateInstruction [User]

A structure that specifies the user or group to add or revoke the role for.