Network.AWS.Grafana.Types.WorkspaceDescription

Description

Documentation

data WorkspaceDescription

A structure containing information about an Amazon Managed Grafana workspace in your account.

See: newWorkspaceDescription smart constructor.

Constructors

WorkspaceDescription'

Fields

workspaceRoleArn :: Maybe (Sensitive Text)

The IAM role that grants permissions to the Amazon Web Services resources that the workspace will view data from. This role must already exist.

freeTrialExpiration :: Maybe POSIX

If this workspace is currently in the free trial period for Grafana Enterprise, this value specifies when that free trial ends.

licenseType :: Maybe LicenseType

Specifies whether this workspace has a full Grafana Enterprise license or a free trial license.

permissionType :: Maybe PermissionType

If this is Service Managed, Amazon Managed Grafana automatically creates the IAM roles and provisions the permissions that the workspace needs to use Amazon Web Services data sources and notification channels.

If this is CUSTOMER_MANAGED, you manage those roles and permissions yourself. If you are creating this workspace in a member account of an organization and that account is not a delegated administrator account, and you want the workspace to access data sources in other Amazon Web Services accounts in the organization, you must choose CUSTOMER_MANAGED.

For more information, see Amazon Managed Grafana permissions and policies for Amazon Web Services data sources and notification channels

name :: Maybe (Sensitive Text)

The name of the workspace.

notificationDestinations :: Maybe [NotificationDestinationType]

The Amazon Web Services notification channels that Amazon Managed Grafana can automatically create IAM roles and permissions for, to allow Amazon Managed Grafana to use these channels.

accountAccessType :: Maybe AccountAccessType

Specifies whether the workspace can access Amazon Web Services resources in this Amazon Web Services account only, or whether it can also access Amazon Web Services resources in other accounts in the same organization. If this is ORGANIZATION, the workspaceOrganizationalUnits parameter specifies which organizational units the workspace can access.

licenseExpiration :: Maybe POSIX

If this workspace has a full Grafana Enterprise license, this specifies when the license ends and will need to be renewed.

organizationRoleName :: Maybe (Sensitive Text)

The name of the IAM role that is used to access resources through Organizations.

stackSetName :: Maybe Text

The name of the CloudFormation stack set that is used to generate IAM roles to be used for this workspace.

organizationalUnits :: Maybe (Sensitive [Text])

Specifies the organizational units that this workspace is allowed to use data sources from, if this workspace is in an account that is part of an organization.

description :: Maybe (Sensitive Text)

The user-defined description of the workspace.

freeTrialConsumed :: Maybe Bool

Specifies whether this workspace has already fully used its free trial for Grafana Enterprise.

authentication :: AuthenticationSummary

A structure that describes whether the workspace uses SAML, Amazon Web Services SSO, or both methods for user authentication.

created :: POSIX

The date that the workspace was created.

dataSources :: [DataSourceType]

Specifies the Amazon Web Services data sources that have been configured to have IAM roles and permissions created to allow Amazon Managed Grafana to read data from these sources.

endpoint :: Text

The URL that users can use to access the Grafana console in the workspace.

grafanaVersion :: Text

The version of Grafana supported in this workspace.

id :: Text

The unique ID of this workspace.

modified :: POSIX

The most recent date that the workspace was modified.

status :: WorkspaceStatus

The current status of the workspace.

newWorkspaceDescription

Arguments

:: AuthenticationSummary
$sel:authentication:WorkspaceDescription'

-> UTCTime
$sel:created:WorkspaceDescription'

-> Text
$sel:endpoint:WorkspaceDescription'

-> Text
$sel:grafanaVersion:WorkspaceDescription'

-> Text
$sel:id:WorkspaceDescription'

-> UTCTime
$sel:modified:WorkspaceDescription'

-> WorkspaceStatus
$sel:status:WorkspaceDescription'

-> WorkspaceDescription

Create a value of WorkspaceDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workspaceRoleArn:WorkspaceDescription', workspaceDescription_workspaceRoleArn - The IAM role that grants permissions to the Amazon Web Services resources that the workspace will view data from. This role must already exist.

$sel:freeTrialExpiration:WorkspaceDescription', workspaceDescription_freeTrialExpiration - If this workspace is currently in the free trial period for Grafana Enterprise, this value specifies when that free trial ends.

$sel:licenseType:WorkspaceDescription', workspaceDescription_licenseType - Specifies whether this workspace has a full Grafana Enterprise license or a free trial license.

$sel:permissionType:WorkspaceDescription', workspaceDescription_permissionType - If this is Service Managed, Amazon Managed Grafana automatically creates the IAM roles and provisions the permissions that the workspace needs to use Amazon Web Services data sources and notification channels.

If this is CUSTOMER_MANAGED, you manage those roles and permissions yourself. If you are creating this workspace in a member account of an organization and that account is not a delegated administrator account, and you want the workspace to access data sources in other Amazon Web Services accounts in the organization, you must choose CUSTOMER_MANAGED.

For more information, see Amazon Managed Grafana permissions and policies for Amazon Web Services data sources and notification channels

$sel:name:WorkspaceDescription', workspaceDescription_name - The name of the workspace.

$sel:notificationDestinations:WorkspaceDescription', workspaceDescription_notificationDestinations - The Amazon Web Services notification channels that Amazon Managed Grafana can automatically create IAM roles and permissions for, to allow Amazon Managed Grafana to use these channels.

$sel:accountAccessType:WorkspaceDescription', workspaceDescription_accountAccessType - Specifies whether the workspace can access Amazon Web Services resources in this Amazon Web Services account only, or whether it can also access Amazon Web Services resources in other accounts in the same organization. If this is ORGANIZATION, the workspaceOrganizationalUnits parameter specifies which organizational units the workspace can access.

$sel:licenseExpiration:WorkspaceDescription', workspaceDescription_licenseExpiration - If this workspace has a full Grafana Enterprise license, this specifies when the license ends and will need to be renewed.

$sel:organizationRoleName:WorkspaceDescription', workspaceDescription_organizationRoleName - The name of the IAM role that is used to access resources through Organizations.

$sel:stackSetName:WorkspaceDescription', workspaceDescription_stackSetName - The name of the CloudFormation stack set that is used to generate IAM roles to be used for this workspace.

$sel:organizationalUnits:WorkspaceDescription', workspaceDescription_organizationalUnits - Specifies the organizational units that this workspace is allowed to use data sources from, if this workspace is in an account that is part of an organization.

$sel:description:WorkspaceDescription', workspaceDescription_description - The user-defined description of the workspace.

$sel:freeTrialConsumed:WorkspaceDescription', workspaceDescription_freeTrialConsumed - Specifies whether this workspace has already fully used its free trial for Grafana Enterprise.

$sel:authentication:WorkspaceDescription', workspaceDescription_authentication - A structure that describes whether the workspace uses SAML, Amazon Web Services SSO, or both methods for user authentication.

$sel:created:WorkspaceDescription', workspaceDescription_created - The date that the workspace was created.

$sel:dataSources:WorkspaceDescription', workspaceDescription_dataSources - Specifies the Amazon Web Services data sources that have been configured to have IAM roles and permissions created to allow Amazon Managed Grafana to read data from these sources.

$sel:endpoint:WorkspaceDescription', workspaceDescription_endpoint - The URL that users can use to access the Grafana console in the workspace.

$sel:grafanaVersion:WorkspaceDescription', workspaceDescription_grafanaVersion - The version of Grafana supported in this workspace.

$sel:id:WorkspaceDescription', workspaceDescription_id - The unique ID of this workspace.

$sel:modified:WorkspaceDescription', workspaceDescription_modified - The most recent date that the workspace was modified.

$sel:status:WorkspaceDescription', workspaceDescription_status - The current status of the workspace.

workspaceDescription_workspaceRoleArn :: Lens' WorkspaceDescription (Maybe Text)

The IAM role that grants permissions to the Amazon Web Services resources that the workspace will view data from. This role must already exist.

workspaceDescription_freeTrialExpiration :: Lens' WorkspaceDescription (Maybe UTCTime)

If this workspace is currently in the free trial period for Grafana Enterprise, this value specifies when that free trial ends.

workspaceDescription_licenseType :: Lens' WorkspaceDescription (Maybe LicenseType)

Specifies whether this workspace has a full Grafana Enterprise license or a free trial license.

workspaceDescription_permissionType :: Lens' WorkspaceDescription (Maybe PermissionType)

If this is Service Managed, Amazon Managed Grafana automatically creates the IAM roles and provisions the permissions that the workspace needs to use Amazon Web Services data sources and notification channels.

If this is CUSTOMER_MANAGED, you manage those roles and permissions yourself. If you are creating this workspace in a member account of an organization and that account is not a delegated administrator account, and you want the workspace to access data sources in other Amazon Web Services accounts in the organization, you must choose CUSTOMER_MANAGED.

For more information, see Amazon Managed Grafana permissions and policies for Amazon Web Services data sources and notification channels

workspaceDescription_name :: Lens' WorkspaceDescription (Maybe Text)

The name of the workspace.

workspaceDescription_notificationDestinations :: Lens' WorkspaceDescription (Maybe [NotificationDestinationType])

The Amazon Web Services notification channels that Amazon Managed Grafana can automatically create IAM roles and permissions for, to allow Amazon Managed Grafana to use these channels.

workspaceDescription_accountAccessType :: Lens' WorkspaceDescription (Maybe AccountAccessType)

Specifies whether the workspace can access Amazon Web Services resources in this Amazon Web Services account only, or whether it can also access Amazon Web Services resources in other accounts in the same organization. If this is ORGANIZATION, the workspaceOrganizationalUnits parameter specifies which organizational units the workspace can access.

workspaceDescription_licenseExpiration :: Lens' WorkspaceDescription (Maybe UTCTime)

If this workspace has a full Grafana Enterprise license, this specifies when the license ends and will need to be renewed.

workspaceDescription_organizationRoleName :: Lens' WorkspaceDescription (Maybe Text)

The name of the IAM role that is used to access resources through Organizations.

workspaceDescription_stackSetName :: Lens' WorkspaceDescription (Maybe Text)

The name of the CloudFormation stack set that is used to generate IAM roles to be used for this workspace.

workspaceDescription_organizationalUnits :: Lens' WorkspaceDescription (Maybe [Text])

Specifies the organizational units that this workspace is allowed to use data sources from, if this workspace is in an account that is part of an organization.

workspaceDescription_description :: Lens' WorkspaceDescription (Maybe Text)

The user-defined description of the workspace.

workspaceDescription_freeTrialConsumed :: Lens' WorkspaceDescription (Maybe Bool)

Specifies whether this workspace has already fully used its free trial for Grafana Enterprise.

workspaceDescription_authentication :: Lens' WorkspaceDescription AuthenticationSummary

A structure that describes whether the workspace uses SAML, Amazon Web Services SSO, or both methods for user authentication.

workspaceDescription_created :: Lens' WorkspaceDescription UTCTime

The date that the workspace was created.

workspaceDescription_dataSources :: Lens' WorkspaceDescription [DataSourceType]

Specifies the Amazon Web Services data sources that have been configured to have IAM roles and permissions created to allow Amazon Managed Grafana to read data from these sources.

workspaceDescription_endpoint :: Lens' WorkspaceDescription Text

The URL that users can use to access the Grafana console in the workspace.

workspaceDescription_grafanaVersion :: Lens' WorkspaceDescription Text

The version of Grafana supported in this workspace.

workspaceDescription_id :: Lens' WorkspaceDescription Text

The unique ID of this workspace.

workspaceDescription_modified :: Lens' WorkspaceDescription UTCTime

The most recent date that the workspace was modified.

workspaceDescription_status :: Lens' WorkspaceDescription WorkspaceStatus

The current status of the workspace.