Network.AWS.Greengrass.AssociateServiceRoleToAccount

Description

Associates a role with your account. AWS IoT Greengrass will use the role to access your Lambda functions and AWS IoT resources. This is necessary for deployments to succeed. The role must have at least minimum permissions in the policy ''AWSGreengrassResourceAccessRolePolicy''.

Creating a Request

data AssociateServiceRoleToAccount

See: newAssociateServiceRoleToAccount smart constructor.

Constructors

AssociateServiceRoleToAccount'

Fields

roleArn :: Text

The ARN of the service role you wish to associate with your account.

newAssociateServiceRoleToAccount

Arguments

:: Text
$sel:roleArn:AssociateServiceRoleToAccount'

-> AssociateServiceRoleToAccount

Create a value of AssociateServiceRoleToAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleArn:AssociateServiceRoleToAccount', associateServiceRoleToAccount_roleArn - The ARN of the service role you wish to associate with your account.

Request Lenses

associateServiceRoleToAccount_roleArn :: Lens' AssociateServiceRoleToAccount Text

The ARN of the service role you wish to associate with your account.

Destructuring the Response

data AssociateServiceRoleToAccountResponse

See: newAssociateServiceRoleToAccountResponse smart constructor.

Constructors

AssociateServiceRoleToAccountResponse'

Fields

associatedAt :: Maybe Text

The time when the service role was associated with the account.

httpStatus :: Int

The response's http status code.

newAssociateServiceRoleToAccountResponse

Arguments

:: Int
$sel:httpStatus:AssociateServiceRoleToAccountResponse'

-> AssociateServiceRoleToAccountResponse

Create a value of AssociateServiceRoleToAccountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associatedAt:AssociateServiceRoleToAccountResponse', associateServiceRoleToAccountResponse_associatedAt - The time when the service role was associated with the account.

$sel:httpStatus:AssociateServiceRoleToAccountResponse', associateServiceRoleToAccountResponse_httpStatus - The response's http status code.

Response Lenses

associateServiceRoleToAccountResponse_associatedAt :: Lens' AssociateServiceRoleToAccountResponse (Maybe Text)

The time when the service role was associated with the account.

associateServiceRoleToAccountResponse_httpStatus :: Lens' AssociateServiceRoleToAccountResponse Int

The response's http status code.