Network.AWS.Greengrass.CreateConnectorDefinitionVersion

Description

Creates a version of a connector definition which has already been defined.

Creating a Request

data CreateConnectorDefinitionVersion

See: newCreateConnectorDefinitionVersion smart constructor.

Constructors

CreateConnectorDefinitionVersion'

Fields

amznClientToken :: Maybe Text

A client token used to correlate requests and responses.

connectors :: Maybe [Connector]

A list of references to connectors in this version, with their corresponding configuration settings.

connectorDefinitionId :: Text

The ID of the connector definition.

newCreateConnectorDefinitionVersion

Arguments

:: Text
$sel:connectorDefinitionId:CreateConnectorDefinitionVersion'

-> CreateConnectorDefinitionVersion

Create a value of CreateConnectorDefinitionVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:amznClientToken:CreateConnectorDefinitionVersion', createConnectorDefinitionVersion_amznClientToken - A client token used to correlate requests and responses.

$sel:connectors:CreateConnectorDefinitionVersion', createConnectorDefinitionVersion_connectors - A list of references to connectors in this version, with their corresponding configuration settings.

$sel:connectorDefinitionId:CreateConnectorDefinitionVersion', createConnectorDefinitionVersion_connectorDefinitionId - The ID of the connector definition.

Request Lenses

createConnectorDefinitionVersion_amznClientToken :: Lens' CreateConnectorDefinitionVersion (Maybe Text)

A client token used to correlate requests and responses.

createConnectorDefinitionVersion_connectors :: Lens' CreateConnectorDefinitionVersion (Maybe [Connector])

A list of references to connectors in this version, with their corresponding configuration settings.

createConnectorDefinitionVersion_connectorDefinitionId :: Lens' CreateConnectorDefinitionVersion Text

The ID of the connector definition.

Destructuring the Response

data CreateConnectorDefinitionVersionResponse

See: newCreateConnectorDefinitionVersionResponse smart constructor.

Constructors

CreateConnectorDefinitionVersionResponse'

Fields

arn :: Maybe Text

The ARN of the version.

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the version was created.

version :: Maybe Text

The ID of the version.

id :: Maybe Text

The ID of the parent definition that the version is associated with.

httpStatus :: Int

The response's http status code.

newCreateConnectorDefinitionVersionResponse

Arguments

:: Int
$sel:httpStatus:CreateConnectorDefinitionVersionResponse'

-> CreateConnectorDefinitionVersionResponse

Create a value of CreateConnectorDefinitionVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:CreateConnectorDefinitionVersionResponse', createConnectorDefinitionVersionResponse_arn - The ARN of the version.

$sel:creationTimestamp:CreateConnectorDefinitionVersionResponse', createConnectorDefinitionVersionResponse_creationTimestamp - The time, in milliseconds since the epoch, when the version was created.

$sel:version:CreateConnectorDefinitionVersionResponse', createConnectorDefinitionVersionResponse_version - The ID of the version.

$sel:id:CreateConnectorDefinitionVersionResponse', createConnectorDefinitionVersionResponse_id - The ID of the parent definition that the version is associated with.

$sel:httpStatus:CreateConnectorDefinitionVersionResponse', createConnectorDefinitionVersionResponse_httpStatus - The response's http status code.

Response Lenses

createConnectorDefinitionVersionResponse_arn :: Lens' CreateConnectorDefinitionVersionResponse (Maybe Text)

The ARN of the version.

createConnectorDefinitionVersionResponse_creationTimestamp :: Lens' CreateConnectorDefinitionVersionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the version was created.

createConnectorDefinitionVersionResponse_version :: Lens' CreateConnectorDefinitionVersionResponse (Maybe Text)

The ID of the version.

createConnectorDefinitionVersionResponse_id :: Lens' CreateConnectorDefinitionVersionResponse (Maybe Text)

The ID of the parent definition that the version is associated with.

createConnectorDefinitionVersionResponse_httpStatus :: Lens' CreateConnectorDefinitionVersionResponse Int

The response's http status code.