Network.AWS.Greengrass.CreateDeployment

Description

Creates a deployment. ''CreateDeployment'' requests are idempotent with respect to the ''X-Amzn-Client-Token'' token and the request parameters.

Creating a Request

data CreateDeployment

See: newCreateDeployment smart constructor.

Constructors

CreateDeployment'

Fields

deploymentId :: Maybe Text

The ID of the deployment if you wish to redeploy a previous deployment.

groupVersionId :: Maybe Text

The ID of the group version to be deployed.

amznClientToken :: Maybe Text

A client token used to correlate requests and responses.

groupId :: Text

The ID of the Greengrass group.

deploymentType :: DeploymentType

The type of deployment. When used for ''CreateDeployment'', only ''NewDeployment'' and ''Redeployment'' are valid.

newCreateDeployment

Arguments

:: Text
$sel:groupId:CreateDeployment'

-> DeploymentType
$sel:deploymentType:CreateDeployment'

-> CreateDeployment

Create a value of CreateDeployment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentId:CreateDeployment', createDeployment_deploymentId - The ID of the deployment if you wish to redeploy a previous deployment.

$sel:groupVersionId:CreateDeployment', createDeployment_groupVersionId - The ID of the group version to be deployed.

$sel:amznClientToken:CreateDeployment', createDeployment_amznClientToken - A client token used to correlate requests and responses.

$sel:groupId:CreateDeployment', createDeployment_groupId - The ID of the Greengrass group.

$sel:deploymentType:CreateDeployment', createDeployment_deploymentType - The type of deployment. When used for ''CreateDeployment'', only ''NewDeployment'' and ''Redeployment'' are valid.

Request Lenses

createDeployment_deploymentId :: Lens' CreateDeployment (Maybe Text)

The ID of the deployment if you wish to redeploy a previous deployment.

createDeployment_groupVersionId :: Lens' CreateDeployment (Maybe Text)

The ID of the group version to be deployed.

createDeployment_amznClientToken :: Lens' CreateDeployment (Maybe Text)

A client token used to correlate requests and responses.

createDeployment_groupId :: Lens' CreateDeployment Text

The ID of the Greengrass group.

createDeployment_deploymentType :: Lens' CreateDeployment DeploymentType

The type of deployment. When used for ''CreateDeployment'', only ''NewDeployment'' and ''Redeployment'' are valid.

Destructuring the Response

data CreateDeploymentResponse

See: newCreateDeploymentResponse smart constructor.

Constructors

CreateDeploymentResponse'

Fields

deploymentId :: Maybe Text

The ID of the deployment.

deploymentArn :: Maybe Text

The ARN of the deployment.

httpStatus :: Int

The response's http status code.

newCreateDeploymentResponse

Arguments

:: Int
$sel:httpStatus:CreateDeploymentResponse'

-> CreateDeploymentResponse

Create a value of CreateDeploymentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentId:CreateDeployment', createDeploymentResponse_deploymentId - The ID of the deployment.

$sel:deploymentArn:CreateDeploymentResponse', createDeploymentResponse_deploymentArn - The ARN of the deployment.

$sel:httpStatus:CreateDeploymentResponse', createDeploymentResponse_httpStatus - The response's http status code.

Response Lenses

createDeploymentResponse_deploymentId :: Lens' CreateDeploymentResponse (Maybe Text)

The ID of the deployment.

createDeploymentResponse_deploymentArn :: Lens' CreateDeploymentResponse (Maybe Text)

The ARN of the deployment.

createDeploymentResponse_httpStatus :: Lens' CreateDeploymentResponse Int

The response's http status code.