Network.AWS.Greengrass.CreateDeviceDefinition

Description

Creates a device definition. You may provide the initial version of the device definition now or use ''CreateDeviceDefinitionVersion'' at a later time.

Creating a Request

data CreateDeviceDefinition

See: newCreateDeviceDefinition smart constructor.

Constructors

CreateDeviceDefinition'

Fields

amznClientToken :: Maybe Text

A client token used to correlate requests and responses.

initialVersion :: Maybe DeviceDefinitionVersion

Information about the initial version of the device definition.

name :: Maybe Text

The name of the device definition.

tags :: Maybe (HashMap Text Text)

Tag(s) to add to the new resource.

newCreateDeviceDefinition :: CreateDeviceDefinition

Create a value of CreateDeviceDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:amznClientToken:CreateDeviceDefinition', createDeviceDefinition_amznClientToken - A client token used to correlate requests and responses.

$sel:initialVersion:CreateDeviceDefinition', createDeviceDefinition_initialVersion - Information about the initial version of the device definition.

$sel:name:CreateDeviceDefinition', createDeviceDefinition_name - The name of the device definition.

$sel:tags:CreateDeviceDefinition', createDeviceDefinition_tags - Tag(s) to add to the new resource.

Request Lenses

createDeviceDefinition_amznClientToken :: Lens' CreateDeviceDefinition (Maybe Text)

A client token used to correlate requests and responses.

createDeviceDefinition_initialVersion :: Lens' CreateDeviceDefinition (Maybe DeviceDefinitionVersion)

Information about the initial version of the device definition.

createDeviceDefinition_name :: Lens' CreateDeviceDefinition (Maybe Text)

The name of the device definition.

createDeviceDefinition_tags :: Lens' CreateDeviceDefinition (Maybe (HashMap Text Text))

Tag(s) to add to the new resource.

Destructuring the Response

data CreateDeviceDefinitionResponse

See: newCreateDeviceDefinitionResponse smart constructor.

Constructors

CreateDeviceDefinitionResponse'

Fields

latestVersionArn :: Maybe Text

The ARN of the latest version associated with the definition.

arn :: Maybe Text

The ARN of the definition.

name :: Maybe Text

The name of the definition.

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was created.

id :: Maybe Text

The ID of the definition.

latestVersion :: Maybe Text

The ID of the latest version associated with the definition.

lastUpdatedTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was last updated.

httpStatus :: Int

The response's http status code.

newCreateDeviceDefinitionResponse

Arguments

:: Int
$sel:httpStatus:CreateDeviceDefinitionResponse'

-> CreateDeviceDefinitionResponse

Create a value of CreateDeviceDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:latestVersionArn:CreateDeviceDefinitionResponse', createDeviceDefinitionResponse_latestVersionArn - The ARN of the latest version associated with the definition.

$sel:arn:CreateDeviceDefinitionResponse', createDeviceDefinitionResponse_arn - The ARN of the definition.

$sel:name:CreateDeviceDefinition', createDeviceDefinitionResponse_name - The name of the definition.

$sel:creationTimestamp:CreateDeviceDefinitionResponse', createDeviceDefinitionResponse_creationTimestamp - The time, in milliseconds since the epoch, when the definition was created.

$sel:id:CreateDeviceDefinitionResponse', createDeviceDefinitionResponse_id - The ID of the definition.

$sel:latestVersion:CreateDeviceDefinitionResponse', createDeviceDefinitionResponse_latestVersion - The ID of the latest version associated with the definition.

$sel:lastUpdatedTimestamp:CreateDeviceDefinitionResponse', createDeviceDefinitionResponse_lastUpdatedTimestamp - The time, in milliseconds since the epoch, when the definition was last updated.

$sel:httpStatus:CreateDeviceDefinitionResponse', createDeviceDefinitionResponse_httpStatus - The response's http status code.

Response Lenses

createDeviceDefinitionResponse_latestVersionArn :: Lens' CreateDeviceDefinitionResponse (Maybe Text)

The ARN of the latest version associated with the definition.

createDeviceDefinitionResponse_arn :: Lens' CreateDeviceDefinitionResponse (Maybe Text)

The ARN of the definition.

createDeviceDefinitionResponse_name :: Lens' CreateDeviceDefinitionResponse (Maybe Text)

The name of the definition.

createDeviceDefinitionResponse_creationTimestamp :: Lens' CreateDeviceDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was created.

createDeviceDefinitionResponse_id :: Lens' CreateDeviceDefinitionResponse (Maybe Text)

The ID of the definition.

createDeviceDefinitionResponse_latestVersion :: Lens' CreateDeviceDefinitionResponse (Maybe Text)

The ID of the latest version associated with the definition.

createDeviceDefinitionResponse_lastUpdatedTimestamp :: Lens' CreateDeviceDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was last updated.

createDeviceDefinitionResponse_httpStatus :: Lens' CreateDeviceDefinitionResponse Int

The response's http status code.