Network.AWS.Greengrass.CreateFunctionDefinition

Description

Creates a Lambda function definition which contains a list of Lambda functions and their configurations to be used in a group. You can create an initial version of the definition by providing a list of Lambda functions and their configurations now, or use ''CreateFunctionDefinitionVersion'' later.

Creating a Request

data CreateFunctionDefinition

See: newCreateFunctionDefinition smart constructor.

Constructors

CreateFunctionDefinition'

Fields

amznClientToken :: Maybe Text

A client token used to correlate requests and responses.

initialVersion :: Maybe FunctionDefinitionVersion

Information about the initial version of the function definition.

name :: Maybe Text

The name of the function definition.

tags :: Maybe (HashMap Text Text)

Tag(s) to add to the new resource.

newCreateFunctionDefinition :: CreateFunctionDefinition

Create a value of CreateFunctionDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:amznClientToken:CreateFunctionDefinition', createFunctionDefinition_amznClientToken - A client token used to correlate requests and responses.

$sel:initialVersion:CreateFunctionDefinition', createFunctionDefinition_initialVersion - Information about the initial version of the function definition.

$sel:name:CreateFunctionDefinition', createFunctionDefinition_name - The name of the function definition.

$sel:tags:CreateFunctionDefinition', createFunctionDefinition_tags - Tag(s) to add to the new resource.

Request Lenses

createFunctionDefinition_amznClientToken :: Lens' CreateFunctionDefinition (Maybe Text)

A client token used to correlate requests and responses.

createFunctionDefinition_initialVersion :: Lens' CreateFunctionDefinition (Maybe FunctionDefinitionVersion)

Information about the initial version of the function definition.

createFunctionDefinition_name :: Lens' CreateFunctionDefinition (Maybe Text)

The name of the function definition.

createFunctionDefinition_tags :: Lens' CreateFunctionDefinition (Maybe (HashMap Text Text))

Tag(s) to add to the new resource.

Destructuring the Response

data CreateFunctionDefinitionResponse

See: newCreateFunctionDefinitionResponse smart constructor.

Constructors

CreateFunctionDefinitionResponse'

Fields

latestVersionArn :: Maybe Text

The ARN of the latest version associated with the definition.

arn :: Maybe Text

The ARN of the definition.

name :: Maybe Text

The name of the definition.

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was created.

id :: Maybe Text

The ID of the definition.

latestVersion :: Maybe Text

The ID of the latest version associated with the definition.

lastUpdatedTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was last updated.

httpStatus :: Int

The response's http status code.

newCreateFunctionDefinitionResponse

Arguments

:: Int
$sel:httpStatus:CreateFunctionDefinitionResponse'

-> CreateFunctionDefinitionResponse

Create a value of CreateFunctionDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:latestVersionArn:CreateFunctionDefinitionResponse', createFunctionDefinitionResponse_latestVersionArn - The ARN of the latest version associated with the definition.

$sel:arn:CreateFunctionDefinitionResponse', createFunctionDefinitionResponse_arn - The ARN of the definition.

$sel:name:CreateFunctionDefinition', createFunctionDefinitionResponse_name - The name of the definition.

$sel:creationTimestamp:CreateFunctionDefinitionResponse', createFunctionDefinitionResponse_creationTimestamp - The time, in milliseconds since the epoch, when the definition was created.

$sel:id:CreateFunctionDefinitionResponse', createFunctionDefinitionResponse_id - The ID of the definition.

$sel:latestVersion:CreateFunctionDefinitionResponse', createFunctionDefinitionResponse_latestVersion - The ID of the latest version associated with the definition.

$sel:lastUpdatedTimestamp:CreateFunctionDefinitionResponse', createFunctionDefinitionResponse_lastUpdatedTimestamp - The time, in milliseconds since the epoch, when the definition was last updated.

$sel:httpStatus:CreateFunctionDefinitionResponse', createFunctionDefinitionResponse_httpStatus - The response's http status code.

Response Lenses

createFunctionDefinitionResponse_latestVersionArn :: Lens' CreateFunctionDefinitionResponse (Maybe Text)

The ARN of the latest version associated with the definition.

createFunctionDefinitionResponse_arn :: Lens' CreateFunctionDefinitionResponse (Maybe Text)

The ARN of the definition.

createFunctionDefinitionResponse_name :: Lens' CreateFunctionDefinitionResponse (Maybe Text)

The name of the definition.

createFunctionDefinitionResponse_creationTimestamp :: Lens' CreateFunctionDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was created.

createFunctionDefinitionResponse_id :: Lens' CreateFunctionDefinitionResponse (Maybe Text)

The ID of the definition.

createFunctionDefinitionResponse_latestVersion :: Lens' CreateFunctionDefinitionResponse (Maybe Text)

The ID of the latest version associated with the definition.

createFunctionDefinitionResponse_lastUpdatedTimestamp :: Lens' CreateFunctionDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was last updated.

createFunctionDefinitionResponse_httpStatus :: Lens' CreateFunctionDefinitionResponse Int

The response's http status code.