Network.AWS.Greengrass.CreateFunctionDefinitionVersion

Description

Creates a version of a Lambda function definition that has already been defined.

Creating a Request

data CreateFunctionDefinitionVersion

Information needed to create a function definition version.

See: newCreateFunctionDefinitionVersion smart constructor.

Constructors

CreateFunctionDefinitionVersion'

Fields

amznClientToken :: Maybe Text

A client token used to correlate requests and responses.

defaultConfig :: Maybe FunctionDefaultConfig

The default configuration that applies to all Lambda functions in this function definition version. Individual Lambda functions can override these settings.

functions :: Maybe [Function]

A list of Lambda functions in this function definition version.

functionDefinitionId :: Text

The ID of the Lambda function definition.

newCreateFunctionDefinitionVersion

Arguments

:: Text
$sel:functionDefinitionId:CreateFunctionDefinitionVersion'

-> CreateFunctionDefinitionVersion

Create a value of CreateFunctionDefinitionVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:amznClientToken:CreateFunctionDefinitionVersion', createFunctionDefinitionVersion_amznClientToken - A client token used to correlate requests and responses.

$sel:defaultConfig:CreateFunctionDefinitionVersion', createFunctionDefinitionVersion_defaultConfig - The default configuration that applies to all Lambda functions in this function definition version. Individual Lambda functions can override these settings.

$sel:functions:CreateFunctionDefinitionVersion', createFunctionDefinitionVersion_functions - A list of Lambda functions in this function definition version.

$sel:functionDefinitionId:CreateFunctionDefinitionVersion', createFunctionDefinitionVersion_functionDefinitionId - The ID of the Lambda function definition.

Request Lenses

createFunctionDefinitionVersion_amznClientToken :: Lens' CreateFunctionDefinitionVersion (Maybe Text)

A client token used to correlate requests and responses.

createFunctionDefinitionVersion_defaultConfig :: Lens' CreateFunctionDefinitionVersion (Maybe FunctionDefaultConfig)

The default configuration that applies to all Lambda functions in this function definition version. Individual Lambda functions can override these settings.

createFunctionDefinitionVersion_functions :: Lens' CreateFunctionDefinitionVersion (Maybe [Function])

A list of Lambda functions in this function definition version.

createFunctionDefinitionVersion_functionDefinitionId :: Lens' CreateFunctionDefinitionVersion Text

The ID of the Lambda function definition.

Destructuring the Response

data CreateFunctionDefinitionVersionResponse

See: newCreateFunctionDefinitionVersionResponse smart constructor.

Constructors

CreateFunctionDefinitionVersionResponse'

Fields

arn :: Maybe Text

The ARN of the version.

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the version was created.

version :: Maybe Text

The ID of the version.

id :: Maybe Text

The ID of the parent definition that the version is associated with.

httpStatus :: Int

The response's http status code.

newCreateFunctionDefinitionVersionResponse

Arguments

:: Int
$sel:httpStatus:CreateFunctionDefinitionVersionResponse'

-> CreateFunctionDefinitionVersionResponse

Create a value of CreateFunctionDefinitionVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:CreateFunctionDefinitionVersionResponse', createFunctionDefinitionVersionResponse_arn - The ARN of the version.

$sel:creationTimestamp:CreateFunctionDefinitionVersionResponse', createFunctionDefinitionVersionResponse_creationTimestamp - The time, in milliseconds since the epoch, when the version was created.

$sel:version:CreateFunctionDefinitionVersionResponse', createFunctionDefinitionVersionResponse_version - The ID of the version.

$sel:id:CreateFunctionDefinitionVersionResponse', createFunctionDefinitionVersionResponse_id - The ID of the parent definition that the version is associated with.

$sel:httpStatus:CreateFunctionDefinitionVersionResponse', createFunctionDefinitionVersionResponse_httpStatus - The response's http status code.

Response Lenses

createFunctionDefinitionVersionResponse_arn :: Lens' CreateFunctionDefinitionVersionResponse (Maybe Text)

The ARN of the version.

createFunctionDefinitionVersionResponse_creationTimestamp :: Lens' CreateFunctionDefinitionVersionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the version was created.

createFunctionDefinitionVersionResponse_version :: Lens' CreateFunctionDefinitionVersionResponse (Maybe Text)

The ID of the version.

createFunctionDefinitionVersionResponse_id :: Lens' CreateFunctionDefinitionVersionResponse (Maybe Text)

The ID of the parent definition that the version is associated with.

createFunctionDefinitionVersionResponse_httpStatus :: Lens' CreateFunctionDefinitionVersionResponse Int

The response's http status code.