Network.AWS.Greengrass.CreateGroup

Description

Creates a group. You may provide the initial version of the group or use ''CreateGroupVersion'' at a later time. Tip: You can use the ''gg_group_setup'' package (https://github.com/awslabs/aws-greengrass-group-setup) as a library or command-line application to create and deploy Greengrass groups.

Creating a Request

data CreateGroup

See: newCreateGroup smart constructor.

Constructors

CreateGroup'

Fields

initialVersion :: Maybe GroupVersion

Information about the initial version of the group.

tags :: Maybe (HashMap Text Text)

Tag(s) to add to the new resource.

amznClientToken :: Maybe Text

A client token used to correlate requests and responses.

name :: Text

The name of the group.

newCreateGroup

Arguments

:: Text
$sel:name:CreateGroup'

-> CreateGroup

Create a value of CreateGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:initialVersion:CreateGroup', createGroup_initialVersion - Information about the initial version of the group.

$sel:tags:CreateGroup', createGroup_tags - Tag(s) to add to the new resource.

$sel:amznClientToken:CreateGroup', createGroup_amznClientToken - A client token used to correlate requests and responses.

$sel:name:CreateGroup', createGroup_name - The name of the group.

Request Lenses

createGroup_initialVersion :: Lens' CreateGroup (Maybe GroupVersion)

Information about the initial version of the group.

createGroup_tags :: Lens' CreateGroup (Maybe (HashMap Text Text))

Tag(s) to add to the new resource.

createGroup_amznClientToken :: Lens' CreateGroup (Maybe Text)

A client token used to correlate requests and responses.

createGroup_name :: Lens' CreateGroup Text

The name of the group.

Destructuring the Response

data CreateGroupResponse

See: newCreateGroupResponse smart constructor.

Constructors

CreateGroupResponse'

Fields

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was created.

latestVersionArn :: Maybe Text

The ARN of the latest version associated with the definition.

latestVersion :: Maybe Text

The ID of the latest version associated with the definition.

arn :: Maybe Text

The ARN of the definition.

id :: Maybe Text

The ID of the definition.

name :: Maybe Text

The name of the definition.

lastUpdatedTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was last updated.

httpStatus :: Int

The response's http status code.

newCreateGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateGroupResponse'

-> CreateGroupResponse

Create a value of CreateGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTimestamp:CreateGroupResponse', createGroupResponse_creationTimestamp - The time, in milliseconds since the epoch, when the definition was created.

$sel:latestVersionArn:CreateGroupResponse', createGroupResponse_latestVersionArn - The ARN of the latest version associated with the definition.

$sel:latestVersion:CreateGroupResponse', createGroupResponse_latestVersion - The ID of the latest version associated with the definition.

$sel:arn:CreateGroupResponse', createGroupResponse_arn - The ARN of the definition.

$sel:id:CreateGroupResponse', createGroupResponse_id - The ID of the definition.

$sel:name:CreateGroup', createGroupResponse_name - The name of the definition.

$sel:lastUpdatedTimestamp:CreateGroupResponse', createGroupResponse_lastUpdatedTimestamp - The time, in milliseconds since the epoch, when the definition was last updated.

$sel:httpStatus:CreateGroupResponse', createGroupResponse_httpStatus - The response's http status code.

Response Lenses

createGroupResponse_creationTimestamp :: Lens' CreateGroupResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was created.

createGroupResponse_latestVersionArn :: Lens' CreateGroupResponse (Maybe Text)

The ARN of the latest version associated with the definition.

createGroupResponse_latestVersion :: Lens' CreateGroupResponse (Maybe Text)

The ID of the latest version associated with the definition.

createGroupResponse_arn :: Lens' CreateGroupResponse (Maybe Text)

The ARN of the definition.

createGroupResponse_id :: Lens' CreateGroupResponse (Maybe Text)

The ID of the definition.

createGroupResponse_name :: Lens' CreateGroupResponse (Maybe Text)

The name of the definition.

createGroupResponse_lastUpdatedTimestamp :: Lens' CreateGroupResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was last updated.

createGroupResponse_httpStatus :: Lens' CreateGroupResponse Int

The response's http status code.