Network.AWS.Greengrass.CreateResourceDefinition

Description

Creates a resource definition which contains a list of resources to be used in a group. You can create an initial version of the definition by providing a list of resources now, or use ''CreateResourceDefinitionVersion'' later.

Creating a Request

data CreateResourceDefinition

See: newCreateResourceDefinition smart constructor.

Constructors

CreateResourceDefinition'

Fields

name :: Maybe Text

The name of the resource definition.

initialVersion :: Maybe ResourceDefinitionVersion

Information about the initial version of the resource definition.

tags :: Maybe (HashMap Text Text)

Tag(s) to add to the new resource.

amznClientToken :: Maybe Text

A client token used to correlate requests and responses.

newCreateResourceDefinition :: CreateResourceDefinition

Create a value of CreateResourceDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:CreateResourceDefinition', createResourceDefinition_name - The name of the resource definition.

$sel:initialVersion:CreateResourceDefinition', createResourceDefinition_initialVersion - Information about the initial version of the resource definition.

$sel:tags:CreateResourceDefinition', createResourceDefinition_tags - Tag(s) to add to the new resource.

$sel:amznClientToken:CreateResourceDefinition', createResourceDefinition_amznClientToken - A client token used to correlate requests and responses.

Request Lenses

createResourceDefinition_name :: Lens' CreateResourceDefinition (Maybe Text)

The name of the resource definition.

createResourceDefinition_initialVersion :: Lens' CreateResourceDefinition (Maybe ResourceDefinitionVersion)

Information about the initial version of the resource definition.

createResourceDefinition_tags :: Lens' CreateResourceDefinition (Maybe (HashMap Text Text))

Tag(s) to add to the new resource.

createResourceDefinition_amznClientToken :: Lens' CreateResourceDefinition (Maybe Text)

A client token used to correlate requests and responses.

Destructuring the Response

data CreateResourceDefinitionResponse

See: newCreateResourceDefinitionResponse smart constructor.

Constructors

CreateResourceDefinitionResponse'

Fields

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was created.

latestVersionArn :: Maybe Text

The ARN of the latest version associated with the definition.

latestVersion :: Maybe Text

The ID of the latest version associated with the definition.

arn :: Maybe Text

The ARN of the definition.

id :: Maybe Text

The ID of the definition.

name :: Maybe Text

The name of the definition.

lastUpdatedTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was last updated.

httpStatus :: Int

The response's http status code.

newCreateResourceDefinitionResponse

Arguments

:: Int
$sel:httpStatus:CreateResourceDefinitionResponse'

-> CreateResourceDefinitionResponse

Create a value of CreateResourceDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTimestamp:CreateResourceDefinitionResponse', createResourceDefinitionResponse_creationTimestamp - The time, in milliseconds since the epoch, when the definition was created.

$sel:latestVersionArn:CreateResourceDefinitionResponse', createResourceDefinitionResponse_latestVersionArn - The ARN of the latest version associated with the definition.

$sel:latestVersion:CreateResourceDefinitionResponse', createResourceDefinitionResponse_latestVersion - The ID of the latest version associated with the definition.

$sel:arn:CreateResourceDefinitionResponse', createResourceDefinitionResponse_arn - The ARN of the definition.

$sel:id:CreateResourceDefinitionResponse', createResourceDefinitionResponse_id - The ID of the definition.

$sel:name:CreateResourceDefinition', createResourceDefinitionResponse_name - The name of the definition.

$sel:lastUpdatedTimestamp:CreateResourceDefinitionResponse', createResourceDefinitionResponse_lastUpdatedTimestamp - The time, in milliseconds since the epoch, when the definition was last updated.

$sel:httpStatus:CreateResourceDefinitionResponse', createResourceDefinitionResponse_httpStatus - The response's http status code.

Response Lenses

createResourceDefinitionResponse_creationTimestamp :: Lens' CreateResourceDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was created.

createResourceDefinitionResponse_latestVersionArn :: Lens' CreateResourceDefinitionResponse (Maybe Text)

The ARN of the latest version associated with the definition.

createResourceDefinitionResponse_latestVersion :: Lens' CreateResourceDefinitionResponse (Maybe Text)

The ID of the latest version associated with the definition.

createResourceDefinitionResponse_arn :: Lens' CreateResourceDefinitionResponse (Maybe Text)

The ARN of the definition.

createResourceDefinitionResponse_id :: Lens' CreateResourceDefinitionResponse (Maybe Text)

The ID of the definition.

createResourceDefinitionResponse_name :: Lens' CreateResourceDefinitionResponse (Maybe Text)

The name of the definition.

createResourceDefinitionResponse_lastUpdatedTimestamp :: Lens' CreateResourceDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was last updated.

createResourceDefinitionResponse_httpStatus :: Lens' CreateResourceDefinitionResponse Int

The response's http status code.