Network.AWS.Greengrass.CreateResourceDefinitionVersion

Description

Creates a version of a resource definition that has already been defined.

Creating a Request

data CreateResourceDefinitionVersion

See: newCreateResourceDefinitionVersion smart constructor.

Constructors

CreateResourceDefinitionVersion'

Fields

amznClientToken :: Maybe Text

A client token used to correlate requests and responses.

resources :: Maybe [Resource]

A list of resources.

resourceDefinitionId :: Text

The ID of the resource definition.

newCreateResourceDefinitionVersion

Arguments

:: Text
$sel:resourceDefinitionId:CreateResourceDefinitionVersion'

-> CreateResourceDefinitionVersion

Create a value of CreateResourceDefinitionVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:amznClientToken:CreateResourceDefinitionVersion', createResourceDefinitionVersion_amznClientToken - A client token used to correlate requests and responses.

$sel:resources:CreateResourceDefinitionVersion', createResourceDefinitionVersion_resources - A list of resources.

$sel:resourceDefinitionId:CreateResourceDefinitionVersion', createResourceDefinitionVersion_resourceDefinitionId - The ID of the resource definition.

Request Lenses

createResourceDefinitionVersion_amznClientToken :: Lens' CreateResourceDefinitionVersion (Maybe Text)

A client token used to correlate requests and responses.

createResourceDefinitionVersion_resources :: Lens' CreateResourceDefinitionVersion (Maybe [Resource])

A list of resources.

createResourceDefinitionVersion_resourceDefinitionId :: Lens' CreateResourceDefinitionVersion Text

The ID of the resource definition.

Destructuring the Response

data CreateResourceDefinitionVersionResponse

See: newCreateResourceDefinitionVersionResponse smart constructor.

Constructors

CreateResourceDefinitionVersionResponse'

Fields

arn :: Maybe Text

The ARN of the version.

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the version was created.

version :: Maybe Text

The ID of the version.

id :: Maybe Text

The ID of the parent definition that the version is associated with.

httpStatus :: Int

The response's http status code.

newCreateResourceDefinitionVersionResponse

Arguments

:: Int
$sel:httpStatus:CreateResourceDefinitionVersionResponse'

-> CreateResourceDefinitionVersionResponse

Create a value of CreateResourceDefinitionVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:CreateResourceDefinitionVersionResponse', createResourceDefinitionVersionResponse_arn - The ARN of the version.

$sel:creationTimestamp:CreateResourceDefinitionVersionResponse', createResourceDefinitionVersionResponse_creationTimestamp - The time, in milliseconds since the epoch, when the version was created.

$sel:version:CreateResourceDefinitionVersionResponse', createResourceDefinitionVersionResponse_version - The ID of the version.

$sel:id:CreateResourceDefinitionVersionResponse', createResourceDefinitionVersionResponse_id - The ID of the parent definition that the version is associated with.

$sel:httpStatus:CreateResourceDefinitionVersionResponse', createResourceDefinitionVersionResponse_httpStatus - The response's http status code.

Response Lenses

createResourceDefinitionVersionResponse_arn :: Lens' CreateResourceDefinitionVersionResponse (Maybe Text)

The ARN of the version.

createResourceDefinitionVersionResponse_creationTimestamp :: Lens' CreateResourceDefinitionVersionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the version was created.

createResourceDefinitionVersionResponse_version :: Lens' CreateResourceDefinitionVersionResponse (Maybe Text)

The ID of the version.

createResourceDefinitionVersionResponse_id :: Lens' CreateResourceDefinitionVersionResponse (Maybe Text)

The ID of the parent definition that the version is associated with.

createResourceDefinitionVersionResponse_httpStatus :: Lens' CreateResourceDefinitionVersionResponse Int

The response's http status code.