Network.AWS.Greengrass.CreateSubscriptionDefinition

Description

Creates a subscription definition. You may provide the initial version of the subscription definition now or use ''CreateSubscriptionDefinitionVersion'' at a later time.

Creating a Request

data CreateSubscriptionDefinition

See: newCreateSubscriptionDefinition smart constructor.

Constructors

CreateSubscriptionDefinition'

Fields

name :: Maybe Text

The name of the subscription definition.

initialVersion :: Maybe SubscriptionDefinitionVersion

Information about the initial version of the subscription definition.

tags :: Maybe (HashMap Text Text)

Tag(s) to add to the new resource.

amznClientToken :: Maybe Text

A client token used to correlate requests and responses.

newCreateSubscriptionDefinition :: CreateSubscriptionDefinition

Create a value of CreateSubscriptionDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:CreateSubscriptionDefinition', createSubscriptionDefinition_name - The name of the subscription definition.

$sel:initialVersion:CreateSubscriptionDefinition', createSubscriptionDefinition_initialVersion - Information about the initial version of the subscription definition.

$sel:tags:CreateSubscriptionDefinition', createSubscriptionDefinition_tags - Tag(s) to add to the new resource.

$sel:amznClientToken:CreateSubscriptionDefinition', createSubscriptionDefinition_amznClientToken - A client token used to correlate requests and responses.

Request Lenses

createSubscriptionDefinition_name :: Lens' CreateSubscriptionDefinition (Maybe Text)

The name of the subscription definition.

createSubscriptionDefinition_initialVersion :: Lens' CreateSubscriptionDefinition (Maybe SubscriptionDefinitionVersion)

Information about the initial version of the subscription definition.

createSubscriptionDefinition_tags :: Lens' CreateSubscriptionDefinition (Maybe (HashMap Text Text))

Tag(s) to add to the new resource.

createSubscriptionDefinition_amznClientToken :: Lens' CreateSubscriptionDefinition (Maybe Text)

A client token used to correlate requests and responses.

Destructuring the Response

data CreateSubscriptionDefinitionResponse

See: newCreateSubscriptionDefinitionResponse smart constructor.

Constructors

CreateSubscriptionDefinitionResponse'

Fields

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was created.

latestVersionArn :: Maybe Text

The ARN of the latest version associated with the definition.

latestVersion :: Maybe Text

The ID of the latest version associated with the definition.

arn :: Maybe Text

The ARN of the definition.

id :: Maybe Text

The ID of the definition.

name :: Maybe Text

The name of the definition.

lastUpdatedTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was last updated.

httpStatus :: Int

The response's http status code.

newCreateSubscriptionDefinitionResponse

Arguments

:: Int
$sel:httpStatus:CreateSubscriptionDefinitionResponse'

-> CreateSubscriptionDefinitionResponse

Create a value of CreateSubscriptionDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTimestamp:CreateSubscriptionDefinitionResponse', createSubscriptionDefinitionResponse_creationTimestamp - The time, in milliseconds since the epoch, when the definition was created.

$sel:latestVersionArn:CreateSubscriptionDefinitionResponse', createSubscriptionDefinitionResponse_latestVersionArn - The ARN of the latest version associated with the definition.

$sel:latestVersion:CreateSubscriptionDefinitionResponse', createSubscriptionDefinitionResponse_latestVersion - The ID of the latest version associated with the definition.

$sel:arn:CreateSubscriptionDefinitionResponse', createSubscriptionDefinitionResponse_arn - The ARN of the definition.

$sel:id:CreateSubscriptionDefinitionResponse', createSubscriptionDefinitionResponse_id - The ID of the definition.

$sel:name:CreateSubscriptionDefinition', createSubscriptionDefinitionResponse_name - The name of the definition.

$sel:lastUpdatedTimestamp:CreateSubscriptionDefinitionResponse', createSubscriptionDefinitionResponse_lastUpdatedTimestamp - The time, in milliseconds since the epoch, when the definition was last updated.

$sel:httpStatus:CreateSubscriptionDefinitionResponse', createSubscriptionDefinitionResponse_httpStatus - The response's http status code.

Response Lenses

createSubscriptionDefinitionResponse_creationTimestamp :: Lens' CreateSubscriptionDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was created.

createSubscriptionDefinitionResponse_latestVersionArn :: Lens' CreateSubscriptionDefinitionResponse (Maybe Text)

The ARN of the latest version associated with the definition.

createSubscriptionDefinitionResponse_latestVersion :: Lens' CreateSubscriptionDefinitionResponse (Maybe Text)

The ID of the latest version associated with the definition.

createSubscriptionDefinitionResponse_arn :: Lens' CreateSubscriptionDefinitionResponse (Maybe Text)

The ARN of the definition.

createSubscriptionDefinitionResponse_id :: Lens' CreateSubscriptionDefinitionResponse (Maybe Text)

The ID of the definition.

createSubscriptionDefinitionResponse_name :: Lens' CreateSubscriptionDefinitionResponse (Maybe Text)

The name of the definition.

createSubscriptionDefinitionResponse_lastUpdatedTimestamp :: Lens' CreateSubscriptionDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was last updated.

createSubscriptionDefinitionResponse_httpStatus :: Lens' CreateSubscriptionDefinitionResponse Int

The response's http status code.