Network.AWS.Greengrass.CreateSubscriptionDefinitionVersion

Description

Creates a version of a subscription definition which has already been defined.

Creating a Request

data CreateSubscriptionDefinitionVersion

See: newCreateSubscriptionDefinitionVersion smart constructor.

Constructors

CreateSubscriptionDefinitionVersion'

Fields

subscriptions :: Maybe [Subscription]

A list of subscriptions.

amznClientToken :: Maybe Text

A client token used to correlate requests and responses.

subscriptionDefinitionId :: Text

The ID of the subscription definition.

newCreateSubscriptionDefinitionVersion

Arguments

:: Text
$sel:subscriptionDefinitionId:CreateSubscriptionDefinitionVersion'

-> CreateSubscriptionDefinitionVersion

Create a value of CreateSubscriptionDefinitionVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subscriptions:CreateSubscriptionDefinitionVersion', createSubscriptionDefinitionVersion_subscriptions - A list of subscriptions.

$sel:amznClientToken:CreateSubscriptionDefinitionVersion', createSubscriptionDefinitionVersion_amznClientToken - A client token used to correlate requests and responses.

$sel:subscriptionDefinitionId:CreateSubscriptionDefinitionVersion', createSubscriptionDefinitionVersion_subscriptionDefinitionId - The ID of the subscription definition.

Request Lenses

createSubscriptionDefinitionVersion_subscriptions :: Lens' CreateSubscriptionDefinitionVersion (Maybe [Subscription])

A list of subscriptions.

createSubscriptionDefinitionVersion_amznClientToken :: Lens' CreateSubscriptionDefinitionVersion (Maybe Text)

A client token used to correlate requests and responses.

createSubscriptionDefinitionVersion_subscriptionDefinitionId :: Lens' CreateSubscriptionDefinitionVersion Text

The ID of the subscription definition.

Destructuring the Response

data CreateSubscriptionDefinitionVersionResponse

See: newCreateSubscriptionDefinitionVersionResponse smart constructor.

Constructors

CreateSubscriptionDefinitionVersionResponse'

Fields

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the version was created.

arn :: Maybe Text

The ARN of the version.

id :: Maybe Text

The ID of the parent definition that the version is associated with.

version :: Maybe Text

The ID of the version.

httpStatus :: Int

The response's http status code.

newCreateSubscriptionDefinitionVersionResponse

Arguments

:: Int
$sel:httpStatus:CreateSubscriptionDefinitionVersionResponse'

-> CreateSubscriptionDefinitionVersionResponse

Create a value of CreateSubscriptionDefinitionVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTimestamp:CreateSubscriptionDefinitionVersionResponse', createSubscriptionDefinitionVersionResponse_creationTimestamp - The time, in milliseconds since the epoch, when the version was created.

$sel:arn:CreateSubscriptionDefinitionVersionResponse', createSubscriptionDefinitionVersionResponse_arn - The ARN of the version.

$sel:id:CreateSubscriptionDefinitionVersionResponse', createSubscriptionDefinitionVersionResponse_id - The ID of the parent definition that the version is associated with.

$sel:version:CreateSubscriptionDefinitionVersionResponse', createSubscriptionDefinitionVersionResponse_version - The ID of the version.

$sel:httpStatus:CreateSubscriptionDefinitionVersionResponse', createSubscriptionDefinitionVersionResponse_httpStatus - The response's http status code.

Response Lenses

createSubscriptionDefinitionVersionResponse_creationTimestamp :: Lens' CreateSubscriptionDefinitionVersionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the version was created.

createSubscriptionDefinitionVersionResponse_arn :: Lens' CreateSubscriptionDefinitionVersionResponse (Maybe Text)

The ARN of the version.

createSubscriptionDefinitionVersionResponse_id :: Lens' CreateSubscriptionDefinitionVersionResponse (Maybe Text)

The ID of the parent definition that the version is associated with.

createSubscriptionDefinitionVersionResponse_version :: Lens' CreateSubscriptionDefinitionVersionResponse (Maybe Text)

The ID of the version.

createSubscriptionDefinitionVersionResponse_httpStatus :: Lens' CreateSubscriptionDefinitionVersionResponse Int

The response's http status code.