Network.AWS.Greengrass.DisassociateServiceRoleFromAccount

Description

Disassociates the service role from your account. Without a service role, deployments will not work.

Creating a Request

data DisassociateServiceRoleFromAccount

See: newDisassociateServiceRoleFromAccount smart constructor.

Constructors

DisassociateServiceRoleFromAccount'

newDisassociateServiceRoleFromAccount :: DisassociateServiceRoleFromAccount

Create a value of DisassociateServiceRoleFromAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data DisassociateServiceRoleFromAccountResponse

See: newDisassociateServiceRoleFromAccountResponse smart constructor.

Constructors

DisassociateServiceRoleFromAccountResponse'

Fields

disassociatedAt :: Maybe Text

The time when the service role was disassociated from the account.

httpStatus :: Int

The response's http status code.

newDisassociateServiceRoleFromAccountResponse

Arguments

:: Int
$sel:httpStatus:DisassociateServiceRoleFromAccountResponse'

-> DisassociateServiceRoleFromAccountResponse

Create a value of DisassociateServiceRoleFromAccountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:disassociatedAt:DisassociateServiceRoleFromAccountResponse', disassociateServiceRoleFromAccountResponse_disassociatedAt - The time when the service role was disassociated from the account.

$sel:httpStatus:DisassociateServiceRoleFromAccountResponse', disassociateServiceRoleFromAccountResponse_httpStatus - The response's http status code.

Response Lenses

disassociateServiceRoleFromAccountResponse_disassociatedAt :: Lens' DisassociateServiceRoleFromAccountResponse (Maybe Text)

The time when the service role was disassociated from the account.

disassociateServiceRoleFromAccountResponse_httpStatus :: Lens' DisassociateServiceRoleFromAccountResponse Int

The response's http status code.