Network.AWS.Greengrass.GetBulkDeploymentStatus

Description

Returns the status of a bulk deployment.

Creating a Request

data GetBulkDeploymentStatus

See: newGetBulkDeploymentStatus smart constructor.

Constructors

GetBulkDeploymentStatus'

Fields

bulkDeploymentId :: Text

The ID of the bulk deployment.

newGetBulkDeploymentStatus

Arguments

:: Text
$sel:bulkDeploymentId:GetBulkDeploymentStatus'

-> GetBulkDeploymentStatus

Create a value of GetBulkDeploymentStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bulkDeploymentId:GetBulkDeploymentStatus', getBulkDeploymentStatus_bulkDeploymentId - The ID of the bulk deployment.

Request Lenses

getBulkDeploymentStatus_bulkDeploymentId :: Lens' GetBulkDeploymentStatus Text

The ID of the bulk deployment.

Destructuring the Response

data GetBulkDeploymentStatusResponse

See: newGetBulkDeploymentStatusResponse smart constructor.

Constructors

GetBulkDeploymentStatusResponse'

Fields

createdAt :: Maybe Text

The time, in ISO format, when the deployment was created.

errorDetails :: Maybe [ErrorDetail]

Error details

bulkDeploymentStatus :: Maybe BulkDeploymentStatus

The status of the bulk deployment.

errorMessage :: Maybe Text

Error message

bulkDeploymentMetrics :: Maybe BulkDeploymentMetrics

Relevant metrics on input records processed during bulk deployment.

tags :: Maybe (HashMap Text Text)

Tag(s) attached to the resource arn.

httpStatus :: Int

The response's http status code.

newGetBulkDeploymentStatusResponse

Arguments

:: Int
$sel:httpStatus:GetBulkDeploymentStatusResponse'

-> GetBulkDeploymentStatusResponse

Create a value of GetBulkDeploymentStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAt:GetBulkDeploymentStatusResponse', getBulkDeploymentStatusResponse_createdAt - The time, in ISO format, when the deployment was created.

$sel:errorDetails:GetBulkDeploymentStatusResponse', getBulkDeploymentStatusResponse_errorDetails - Error details

$sel:bulkDeploymentStatus:GetBulkDeploymentStatusResponse', getBulkDeploymentStatusResponse_bulkDeploymentStatus - The status of the bulk deployment.

$sel:errorMessage:GetBulkDeploymentStatusResponse', getBulkDeploymentStatusResponse_errorMessage - Error message

$sel:bulkDeploymentMetrics:GetBulkDeploymentStatusResponse', getBulkDeploymentStatusResponse_bulkDeploymentMetrics - Relevant metrics on input records processed during bulk deployment.

$sel:tags:GetBulkDeploymentStatusResponse', getBulkDeploymentStatusResponse_tags - Tag(s) attached to the resource arn.

$sel:httpStatus:GetBulkDeploymentStatusResponse', getBulkDeploymentStatusResponse_httpStatus - The response's http status code.

Response Lenses

getBulkDeploymentStatusResponse_createdAt :: Lens' GetBulkDeploymentStatusResponse (Maybe Text)

The time, in ISO format, when the deployment was created.

getBulkDeploymentStatusResponse_errorDetails :: Lens' GetBulkDeploymentStatusResponse (Maybe [ErrorDetail])

Error details

getBulkDeploymentStatusResponse_bulkDeploymentStatus :: Lens' GetBulkDeploymentStatusResponse (Maybe BulkDeploymentStatus)

The status of the bulk deployment.

getBulkDeploymentStatusResponse_errorMessage :: Lens' GetBulkDeploymentStatusResponse (Maybe Text)

Error message

getBulkDeploymentStatusResponse_bulkDeploymentMetrics :: Lens' GetBulkDeploymentStatusResponse (Maybe BulkDeploymentMetrics)

Relevant metrics on input records processed during bulk deployment.

getBulkDeploymentStatusResponse_tags :: Lens' GetBulkDeploymentStatusResponse (Maybe (HashMap Text Text))

Tag(s) attached to the resource arn.

getBulkDeploymentStatusResponse_httpStatus :: Lens' GetBulkDeploymentStatusResponse Int

The response's http status code.