Network.AWS.Greengrass.GetConnectorDefinition

Description

Retrieves information about a connector definition.

Creating a Request

data GetConnectorDefinition

See: newGetConnectorDefinition smart constructor.

Constructors

GetConnectorDefinition'

Fields

connectorDefinitionId :: Text

The ID of the connector definition.

newGetConnectorDefinition

Arguments

:: Text
$sel:connectorDefinitionId:GetConnectorDefinition'

-> GetConnectorDefinition

Create a value of GetConnectorDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectorDefinitionId:GetConnectorDefinition', getConnectorDefinition_connectorDefinitionId - The ID of the connector definition.

Request Lenses

getConnectorDefinition_connectorDefinitionId :: Lens' GetConnectorDefinition Text

The ID of the connector definition.

Destructuring the Response

data GetConnectorDefinitionResponse

See: newGetConnectorDefinitionResponse smart constructor.

Constructors

GetConnectorDefinitionResponse'

Fields

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was created.

latestVersionArn :: Maybe Text

The ARN of the latest version associated with the definition.

latestVersion :: Maybe Text

The ID of the latest version associated with the definition.

arn :: Maybe Text

The ARN of the definition.

id :: Maybe Text

The ID of the definition.

name :: Maybe Text

The name of the definition.

tags :: Maybe (HashMap Text Text)

Tag(s) attached to the resource arn.

lastUpdatedTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was last updated.

httpStatus :: Int

The response's http status code.

newGetConnectorDefinitionResponse

Arguments

:: Int
$sel:httpStatus:GetConnectorDefinitionResponse'

-> GetConnectorDefinitionResponse

Create a value of GetConnectorDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTimestamp:GetConnectorDefinitionResponse', getConnectorDefinitionResponse_creationTimestamp - The time, in milliseconds since the epoch, when the definition was created.

$sel:latestVersionArn:GetConnectorDefinitionResponse', getConnectorDefinitionResponse_latestVersionArn - The ARN of the latest version associated with the definition.

$sel:latestVersion:GetConnectorDefinitionResponse', getConnectorDefinitionResponse_latestVersion - The ID of the latest version associated with the definition.

$sel:arn:GetConnectorDefinitionResponse', getConnectorDefinitionResponse_arn - The ARN of the definition.

$sel:id:GetConnectorDefinitionResponse', getConnectorDefinitionResponse_id - The ID of the definition.

$sel:name:GetConnectorDefinitionResponse', getConnectorDefinitionResponse_name - The name of the definition.

$sel:tags:GetConnectorDefinitionResponse', getConnectorDefinitionResponse_tags - Tag(s) attached to the resource arn.

$sel:lastUpdatedTimestamp:GetConnectorDefinitionResponse', getConnectorDefinitionResponse_lastUpdatedTimestamp - The time, in milliseconds since the epoch, when the definition was last updated.

$sel:httpStatus:GetConnectorDefinitionResponse', getConnectorDefinitionResponse_httpStatus - The response's http status code.

Response Lenses

getConnectorDefinitionResponse_creationTimestamp :: Lens' GetConnectorDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was created.

getConnectorDefinitionResponse_latestVersionArn :: Lens' GetConnectorDefinitionResponse (Maybe Text)

The ARN of the latest version associated with the definition.

getConnectorDefinitionResponse_latestVersion :: Lens' GetConnectorDefinitionResponse (Maybe Text)

The ID of the latest version associated with the definition.

getConnectorDefinitionResponse_arn :: Lens' GetConnectorDefinitionResponse (Maybe Text)

The ARN of the definition.

getConnectorDefinitionResponse_id :: Lens' GetConnectorDefinitionResponse (Maybe Text)

The ID of the definition.

getConnectorDefinitionResponse_name :: Lens' GetConnectorDefinitionResponse (Maybe Text)

The name of the definition.

getConnectorDefinitionResponse_tags :: Lens' GetConnectorDefinitionResponse (Maybe (HashMap Text Text))

Tag(s) attached to the resource arn.

getConnectorDefinitionResponse_lastUpdatedTimestamp :: Lens' GetConnectorDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was last updated.

getConnectorDefinitionResponse_httpStatus :: Lens' GetConnectorDefinitionResponse Int

The response's http status code.