Network.AWS.Greengrass.GetCoreDefinitionVersion

Description

Retrieves information about a core definition version.

Creating a Request

data GetCoreDefinitionVersion

See: newGetCoreDefinitionVersion smart constructor.

Constructors

GetCoreDefinitionVersion'

Fields

coreDefinitionId :: Text

The ID of the core definition.

coreDefinitionVersionId :: Text

The ID of the core definition version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListCoreDefinitionVersions'' requests. If the version is the last one that was associated with a core definition, the value also maps to the ''LatestVersion'' property of the corresponding ''DefinitionInformation'' object.

newGetCoreDefinitionVersion

Arguments

:: Text
$sel:coreDefinitionId:GetCoreDefinitionVersion'

-> Text
$sel:coreDefinitionVersionId:GetCoreDefinitionVersion'

-> GetCoreDefinitionVersion

Create a value of GetCoreDefinitionVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:coreDefinitionId:GetCoreDefinitionVersion', getCoreDefinitionVersion_coreDefinitionId - The ID of the core definition.

$sel:coreDefinitionVersionId:GetCoreDefinitionVersion', getCoreDefinitionVersion_coreDefinitionVersionId - The ID of the core definition version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListCoreDefinitionVersions'' requests. If the version is the last one that was associated with a core definition, the value also maps to the ''LatestVersion'' property of the corresponding ''DefinitionInformation'' object.

Request Lenses

getCoreDefinitionVersion_coreDefinitionId :: Lens' GetCoreDefinitionVersion Text

The ID of the core definition.

getCoreDefinitionVersion_coreDefinitionVersionId :: Lens' GetCoreDefinitionVersion Text

The ID of the core definition version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListCoreDefinitionVersions'' requests. If the version is the last one that was associated with a core definition, the value also maps to the ''LatestVersion'' property of the corresponding ''DefinitionInformation'' object.

Destructuring the Response

data GetCoreDefinitionVersionResponse

See: newGetCoreDefinitionVersionResponse smart constructor.

Constructors

GetCoreDefinitionVersionResponse'

Fields

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the core definition version was created.

nextToken :: Maybe Text

The token for the next set of results, or ''null'' if there are no additional results.

arn :: Maybe Text

The ARN of the core definition version.

id :: Maybe Text

The ID of the core definition version.

version :: Maybe Text

The version of the core definition version.

definition :: Maybe CoreDefinitionVersion

Information about the core definition version.

httpStatus :: Int

The response's http status code.

newGetCoreDefinitionVersionResponse

Arguments

:: Int
$sel:httpStatus:GetCoreDefinitionVersionResponse'

-> GetCoreDefinitionVersionResponse

Create a value of GetCoreDefinitionVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTimestamp:GetCoreDefinitionVersionResponse', getCoreDefinitionVersionResponse_creationTimestamp - The time, in milliseconds since the epoch, when the core definition version was created.

$sel:nextToken:GetCoreDefinitionVersionResponse', getCoreDefinitionVersionResponse_nextToken - The token for the next set of results, or ''null'' if there are no additional results.

$sel:arn:GetCoreDefinitionVersionResponse', getCoreDefinitionVersionResponse_arn - The ARN of the core definition version.

$sel:id:GetCoreDefinitionVersionResponse', getCoreDefinitionVersionResponse_id - The ID of the core definition version.

$sel:version:GetCoreDefinitionVersionResponse', getCoreDefinitionVersionResponse_version - The version of the core definition version.

$sel:definition:GetCoreDefinitionVersionResponse', getCoreDefinitionVersionResponse_definition - Information about the core definition version.

$sel:httpStatus:GetCoreDefinitionVersionResponse', getCoreDefinitionVersionResponse_httpStatus - The response's http status code.

Response Lenses

getCoreDefinitionVersionResponse_creationTimestamp :: Lens' GetCoreDefinitionVersionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the core definition version was created.

getCoreDefinitionVersionResponse_nextToken :: Lens' GetCoreDefinitionVersionResponse (Maybe Text)

The token for the next set of results, or ''null'' if there are no additional results.

getCoreDefinitionVersionResponse_arn :: Lens' GetCoreDefinitionVersionResponse (Maybe Text)

The ARN of the core definition version.

getCoreDefinitionVersionResponse_id :: Lens' GetCoreDefinitionVersionResponse (Maybe Text)

The ID of the core definition version.

getCoreDefinitionVersionResponse_version :: Lens' GetCoreDefinitionVersionResponse (Maybe Text)

The version of the core definition version.

getCoreDefinitionVersionResponse_definition :: Lens' GetCoreDefinitionVersionResponse (Maybe CoreDefinitionVersion)

Information about the core definition version.

getCoreDefinitionVersionResponse_httpStatus :: Lens' GetCoreDefinitionVersionResponse Int

The response's http status code.