Network.AWS.Greengrass.GetDeploymentStatus

Description

Returns the status of a deployment.

Creating a Request

data GetDeploymentStatus

See: newGetDeploymentStatus smart constructor.

Constructors

GetDeploymentStatus'

Fields

groupId :: Text

The ID of the Greengrass group.

deploymentId :: Text

The ID of the deployment.

newGetDeploymentStatus

Arguments

:: Text
$sel:groupId:GetDeploymentStatus'

-> Text
$sel:deploymentId:GetDeploymentStatus'

-> GetDeploymentStatus

Create a value of GetDeploymentStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupId:GetDeploymentStatus', getDeploymentStatus_groupId - The ID of the Greengrass group.

$sel:deploymentId:GetDeploymentStatus', getDeploymentStatus_deploymentId - The ID of the deployment.

Request Lenses

getDeploymentStatus_groupId :: Lens' GetDeploymentStatus Text

The ID of the Greengrass group.

getDeploymentStatus_deploymentId :: Lens' GetDeploymentStatus Text

The ID of the deployment.

Destructuring the Response

data GetDeploymentStatusResponse

See: newGetDeploymentStatusResponse smart constructor.

Constructors

GetDeploymentStatusResponse'

Fields

deploymentType :: Maybe DeploymentType

The type of the deployment.

updatedAt :: Maybe Text

The time, in milliseconds since the epoch, when the deployment status was updated.

deploymentStatus :: Maybe Text

The status of the deployment: ''InProgress'', ''Building'', ''Success'', or ''Failure''.

errorMessage :: Maybe Text

Error message

errorDetails :: Maybe [ErrorDetail]

Error details

httpStatus :: Int

The response's http status code.

newGetDeploymentStatusResponse

Arguments

:: Int
$sel:httpStatus:GetDeploymentStatusResponse'

-> GetDeploymentStatusResponse

Create a value of GetDeploymentStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentType:GetDeploymentStatusResponse', getDeploymentStatusResponse_deploymentType - The type of the deployment.

$sel:updatedAt:GetDeploymentStatusResponse', getDeploymentStatusResponse_updatedAt - The time, in milliseconds since the epoch, when the deployment status was updated.

$sel:deploymentStatus:GetDeploymentStatusResponse', getDeploymentStatusResponse_deploymentStatus - The status of the deployment: ''InProgress'', ''Building'', ''Success'', or ''Failure''.

$sel:errorMessage:GetDeploymentStatusResponse', getDeploymentStatusResponse_errorMessage - Error message

$sel:errorDetails:GetDeploymentStatusResponse', getDeploymentStatusResponse_errorDetails - Error details

$sel:httpStatus:GetDeploymentStatusResponse', getDeploymentStatusResponse_httpStatus - The response's http status code.

Response Lenses

getDeploymentStatusResponse_deploymentType :: Lens' GetDeploymentStatusResponse (Maybe DeploymentType)

The type of the deployment.

getDeploymentStatusResponse_updatedAt :: Lens' GetDeploymentStatusResponse (Maybe Text)

The time, in milliseconds since the epoch, when the deployment status was updated.

getDeploymentStatusResponse_deploymentStatus :: Lens' GetDeploymentStatusResponse (Maybe Text)

The status of the deployment: ''InProgress'', ''Building'', ''Success'', or ''Failure''.

getDeploymentStatusResponse_errorMessage :: Lens' GetDeploymentStatusResponse (Maybe Text)

Error message

getDeploymentStatusResponse_errorDetails :: Lens' GetDeploymentStatusResponse (Maybe [ErrorDetail])

Error details

getDeploymentStatusResponse_httpStatus :: Lens' GetDeploymentStatusResponse Int

The response's http status code.