Network.AWS.Greengrass.GetDeviceDefinition

Description

Retrieves information about a device definition.

Creating a Request

data GetDeviceDefinition

See: newGetDeviceDefinition smart constructor.

Constructors

GetDeviceDefinition'

Fields

deviceDefinitionId :: Text

The ID of the device definition.

newGetDeviceDefinition

Arguments

:: Text
$sel:deviceDefinitionId:GetDeviceDefinition'

-> GetDeviceDefinition

Create a value of GetDeviceDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceDefinitionId:GetDeviceDefinition', getDeviceDefinition_deviceDefinitionId - The ID of the device definition.

Request Lenses

getDeviceDefinition_deviceDefinitionId :: Lens' GetDeviceDefinition Text

The ID of the device definition.

Destructuring the Response

data GetDeviceDefinitionResponse

See: newGetDeviceDefinitionResponse smart constructor.

Constructors

GetDeviceDefinitionResponse'

Fields

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was created.

latestVersionArn :: Maybe Text

The ARN of the latest version associated with the definition.

latestVersion :: Maybe Text

The ID of the latest version associated with the definition.

arn :: Maybe Text

The ARN of the definition.

id :: Maybe Text

The ID of the definition.

name :: Maybe Text

The name of the definition.

tags :: Maybe (HashMap Text Text)

Tag(s) attached to the resource arn.

lastUpdatedTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was last updated.

httpStatus :: Int

The response's http status code.

newGetDeviceDefinitionResponse

Arguments

:: Int
$sel:httpStatus:GetDeviceDefinitionResponse'

-> GetDeviceDefinitionResponse

Create a value of GetDeviceDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTimestamp:GetDeviceDefinitionResponse', getDeviceDefinitionResponse_creationTimestamp - The time, in milliseconds since the epoch, when the definition was created.

$sel:latestVersionArn:GetDeviceDefinitionResponse', getDeviceDefinitionResponse_latestVersionArn - The ARN of the latest version associated with the definition.

$sel:latestVersion:GetDeviceDefinitionResponse', getDeviceDefinitionResponse_latestVersion - The ID of the latest version associated with the definition.

$sel:arn:GetDeviceDefinitionResponse', getDeviceDefinitionResponse_arn - The ARN of the definition.

$sel:id:GetDeviceDefinitionResponse', getDeviceDefinitionResponse_id - The ID of the definition.

$sel:name:GetDeviceDefinitionResponse', getDeviceDefinitionResponse_name - The name of the definition.

$sel:tags:GetDeviceDefinitionResponse', getDeviceDefinitionResponse_tags - Tag(s) attached to the resource arn.

$sel:lastUpdatedTimestamp:GetDeviceDefinitionResponse', getDeviceDefinitionResponse_lastUpdatedTimestamp - The time, in milliseconds since the epoch, when the definition was last updated.

$sel:httpStatus:GetDeviceDefinitionResponse', getDeviceDefinitionResponse_httpStatus - The response's http status code.

Response Lenses

getDeviceDefinitionResponse_creationTimestamp :: Lens' GetDeviceDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was created.

getDeviceDefinitionResponse_latestVersionArn :: Lens' GetDeviceDefinitionResponse (Maybe Text)

The ARN of the latest version associated with the definition.

getDeviceDefinitionResponse_latestVersion :: Lens' GetDeviceDefinitionResponse (Maybe Text)

The ID of the latest version associated with the definition.

getDeviceDefinitionResponse_arn :: Lens' GetDeviceDefinitionResponse (Maybe Text)

The ARN of the definition.

getDeviceDefinitionResponse_id :: Lens' GetDeviceDefinitionResponse (Maybe Text)

The ID of the definition.

getDeviceDefinitionResponse_name :: Lens' GetDeviceDefinitionResponse (Maybe Text)

The name of the definition.

getDeviceDefinitionResponse_tags :: Lens' GetDeviceDefinitionResponse (Maybe (HashMap Text Text))

Tag(s) attached to the resource arn.

getDeviceDefinitionResponse_lastUpdatedTimestamp :: Lens' GetDeviceDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was last updated.

getDeviceDefinitionResponse_httpStatus :: Lens' GetDeviceDefinitionResponse Int

The response's http status code.