Network.AWS.Greengrass.GetFunctionDefinition

Description

Retrieves information about a Lambda function definition, including its creation time and latest version.

Creating a Request

data GetFunctionDefinition

See: newGetFunctionDefinition smart constructor.

Constructors

GetFunctionDefinition'

Fields

functionDefinitionId :: Text

The ID of the Lambda function definition.

newGetFunctionDefinition

Arguments

:: Text
$sel:functionDefinitionId:GetFunctionDefinition'

-> GetFunctionDefinition

Create a value of GetFunctionDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:functionDefinitionId:GetFunctionDefinition', getFunctionDefinition_functionDefinitionId - The ID of the Lambda function definition.

Request Lenses

getFunctionDefinition_functionDefinitionId :: Lens' GetFunctionDefinition Text

The ID of the Lambda function definition.

Destructuring the Response

data GetFunctionDefinitionResponse

See: newGetFunctionDefinitionResponse smart constructor.

Constructors

GetFunctionDefinitionResponse'

Fields

latestVersionArn :: Maybe Text

The ARN of the latest version associated with the definition.

arn :: Maybe Text

The ARN of the definition.

name :: Maybe Text

The name of the definition.

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was created.

id :: Maybe Text

The ID of the definition.

latestVersion :: Maybe Text

The ID of the latest version associated with the definition.

lastUpdatedTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was last updated.

tags :: Maybe (HashMap Text Text)

Tag(s) attached to the resource arn.

httpStatus :: Int

The response's http status code.

newGetFunctionDefinitionResponse

Arguments

:: Int
$sel:httpStatus:GetFunctionDefinitionResponse'

-> GetFunctionDefinitionResponse

Create a value of GetFunctionDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:latestVersionArn:GetFunctionDefinitionResponse', getFunctionDefinitionResponse_latestVersionArn - The ARN of the latest version associated with the definition.

$sel:arn:GetFunctionDefinitionResponse', getFunctionDefinitionResponse_arn - The ARN of the definition.

$sel:name:GetFunctionDefinitionResponse', getFunctionDefinitionResponse_name - The name of the definition.

$sel:creationTimestamp:GetFunctionDefinitionResponse', getFunctionDefinitionResponse_creationTimestamp - The time, in milliseconds since the epoch, when the definition was created.

$sel:id:GetFunctionDefinitionResponse', getFunctionDefinitionResponse_id - The ID of the definition.

$sel:latestVersion:GetFunctionDefinitionResponse', getFunctionDefinitionResponse_latestVersion - The ID of the latest version associated with the definition.

$sel:lastUpdatedTimestamp:GetFunctionDefinitionResponse', getFunctionDefinitionResponse_lastUpdatedTimestamp - The time, in milliseconds since the epoch, when the definition was last updated.

$sel:tags:GetFunctionDefinitionResponse', getFunctionDefinitionResponse_tags - Tag(s) attached to the resource arn.

$sel:httpStatus:GetFunctionDefinitionResponse', getFunctionDefinitionResponse_httpStatus - The response's http status code.

Response Lenses

getFunctionDefinitionResponse_latestVersionArn :: Lens' GetFunctionDefinitionResponse (Maybe Text)

The ARN of the latest version associated with the definition.

getFunctionDefinitionResponse_arn :: Lens' GetFunctionDefinitionResponse (Maybe Text)

The ARN of the definition.

getFunctionDefinitionResponse_name :: Lens' GetFunctionDefinitionResponse (Maybe Text)

The name of the definition.

getFunctionDefinitionResponse_creationTimestamp :: Lens' GetFunctionDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was created.

getFunctionDefinitionResponse_id :: Lens' GetFunctionDefinitionResponse (Maybe Text)

The ID of the definition.

getFunctionDefinitionResponse_latestVersion :: Lens' GetFunctionDefinitionResponse (Maybe Text)

The ID of the latest version associated with the definition.

getFunctionDefinitionResponse_lastUpdatedTimestamp :: Lens' GetFunctionDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was last updated.

getFunctionDefinitionResponse_tags :: Lens' GetFunctionDefinitionResponse (Maybe (HashMap Text Text))

Tag(s) attached to the resource arn.

getFunctionDefinitionResponse_httpStatus :: Lens' GetFunctionDefinitionResponse Int

The response's http status code.