Network.AWS.Greengrass.GetGroup

Description

Retrieves information about a group.

Creating a Request

data GetGroup

See: newGetGroup smart constructor.

Constructors

GetGroup'

Fields

groupId :: Text

The ID of the Greengrass group.

newGetGroup

Arguments

:: Text
$sel:groupId:GetGroup'

-> GetGroup

Create a value of GetGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupId:GetGroup', getGroup_groupId - The ID of the Greengrass group.

Request Lenses

getGroup_groupId :: Lens' GetGroup Text

The ID of the Greengrass group.

Destructuring the Response

data GetGroupResponse

See: newGetGroupResponse smart constructor.

Constructors

GetGroupResponse'

Fields

latestVersionArn :: Maybe Text

The ARN of the latest version associated with the definition.

arn :: Maybe Text

The ARN of the definition.

name :: Maybe Text

The name of the definition.

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was created.

id :: Maybe Text

The ID of the definition.

latestVersion :: Maybe Text

The ID of the latest version associated with the definition.

lastUpdatedTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was last updated.

tags :: Maybe (HashMap Text Text)

Tag(s) attached to the resource arn.

httpStatus :: Int

The response's http status code.

newGetGroupResponse

Arguments

:: Int
$sel:httpStatus:GetGroupResponse'

-> GetGroupResponse

Create a value of GetGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:latestVersionArn:GetGroupResponse', getGroupResponse_latestVersionArn - The ARN of the latest version associated with the definition.

$sel:arn:GetGroupResponse', getGroupResponse_arn - The ARN of the definition.

$sel:name:GetGroupResponse', getGroupResponse_name - The name of the definition.

$sel:creationTimestamp:GetGroupResponse', getGroupResponse_creationTimestamp - The time, in milliseconds since the epoch, when the definition was created.

$sel:id:GetGroupResponse', getGroupResponse_id - The ID of the definition.

$sel:latestVersion:GetGroupResponse', getGroupResponse_latestVersion - The ID of the latest version associated with the definition.

$sel:lastUpdatedTimestamp:GetGroupResponse', getGroupResponse_lastUpdatedTimestamp - The time, in milliseconds since the epoch, when the definition was last updated.

$sel:tags:GetGroupResponse', getGroupResponse_tags - Tag(s) attached to the resource arn.

$sel:httpStatus:GetGroupResponse', getGroupResponse_httpStatus - The response's http status code.

Response Lenses

getGroupResponse_latestVersionArn :: Lens' GetGroupResponse (Maybe Text)

The ARN of the latest version associated with the definition.

getGroupResponse_arn :: Lens' GetGroupResponse (Maybe Text)

The ARN of the definition.

getGroupResponse_name :: Lens' GetGroupResponse (Maybe Text)

The name of the definition.

getGroupResponse_creationTimestamp :: Lens' GetGroupResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was created.

getGroupResponse_id :: Lens' GetGroupResponse (Maybe Text)

The ID of the definition.

getGroupResponse_latestVersion :: Lens' GetGroupResponse (Maybe Text)

The ID of the latest version associated with the definition.

getGroupResponse_lastUpdatedTimestamp :: Lens' GetGroupResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was last updated.

getGroupResponse_tags :: Lens' GetGroupResponse (Maybe (HashMap Text Text))

Tag(s) attached to the resource arn.

getGroupResponse_httpStatus :: Lens' GetGroupResponse Int

The response's http status code.