Network.AWS.Greengrass.GetLoggerDefinition

Description

Retrieves information about a logger definition.

Creating a Request

data GetLoggerDefinition

See: newGetLoggerDefinition smart constructor.

Constructors

GetLoggerDefinition'

Fields

loggerDefinitionId :: Text

The ID of the logger definition.

newGetLoggerDefinition

Arguments

:: Text
$sel:loggerDefinitionId:GetLoggerDefinition'

-> GetLoggerDefinition

Create a value of GetLoggerDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loggerDefinitionId:GetLoggerDefinition', getLoggerDefinition_loggerDefinitionId - The ID of the logger definition.

Request Lenses

getLoggerDefinition_loggerDefinitionId :: Lens' GetLoggerDefinition Text

The ID of the logger definition.

Destructuring the Response

data GetLoggerDefinitionResponse

See: newGetLoggerDefinitionResponse smart constructor.

Constructors

GetLoggerDefinitionResponse'

Fields

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was created.

latestVersionArn :: Maybe Text

The ARN of the latest version associated with the definition.

latestVersion :: Maybe Text

The ID of the latest version associated with the definition.

arn :: Maybe Text

The ARN of the definition.

id :: Maybe Text

The ID of the definition.

name :: Maybe Text

The name of the definition.

tags :: Maybe (HashMap Text Text)

Tag(s) attached to the resource arn.

lastUpdatedTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was last updated.

httpStatus :: Int

The response's http status code.

newGetLoggerDefinitionResponse

Arguments

:: Int
$sel:httpStatus:GetLoggerDefinitionResponse'

-> GetLoggerDefinitionResponse

Create a value of GetLoggerDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTimestamp:GetLoggerDefinitionResponse', getLoggerDefinitionResponse_creationTimestamp - The time, in milliseconds since the epoch, when the definition was created.

$sel:latestVersionArn:GetLoggerDefinitionResponse', getLoggerDefinitionResponse_latestVersionArn - The ARN of the latest version associated with the definition.

$sel:latestVersion:GetLoggerDefinitionResponse', getLoggerDefinitionResponse_latestVersion - The ID of the latest version associated with the definition.

$sel:arn:GetLoggerDefinitionResponse', getLoggerDefinitionResponse_arn - The ARN of the definition.

$sel:id:GetLoggerDefinitionResponse', getLoggerDefinitionResponse_id - The ID of the definition.

$sel:name:GetLoggerDefinitionResponse', getLoggerDefinitionResponse_name - The name of the definition.

$sel:tags:GetLoggerDefinitionResponse', getLoggerDefinitionResponse_tags - Tag(s) attached to the resource arn.

$sel:lastUpdatedTimestamp:GetLoggerDefinitionResponse', getLoggerDefinitionResponse_lastUpdatedTimestamp - The time, in milliseconds since the epoch, when the definition was last updated.

$sel:httpStatus:GetLoggerDefinitionResponse', getLoggerDefinitionResponse_httpStatus - The response's http status code.

Response Lenses

getLoggerDefinitionResponse_creationTimestamp :: Lens' GetLoggerDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was created.

getLoggerDefinitionResponse_latestVersionArn :: Lens' GetLoggerDefinitionResponse (Maybe Text)

The ARN of the latest version associated with the definition.

getLoggerDefinitionResponse_latestVersion :: Lens' GetLoggerDefinitionResponse (Maybe Text)

The ID of the latest version associated with the definition.

getLoggerDefinitionResponse_arn :: Lens' GetLoggerDefinitionResponse (Maybe Text)

The ARN of the definition.

getLoggerDefinitionResponse_id :: Lens' GetLoggerDefinitionResponse (Maybe Text)

The ID of the definition.

getLoggerDefinitionResponse_name :: Lens' GetLoggerDefinitionResponse (Maybe Text)

The name of the definition.

getLoggerDefinitionResponse_tags :: Lens' GetLoggerDefinitionResponse (Maybe (HashMap Text Text))

Tag(s) attached to the resource arn.

getLoggerDefinitionResponse_lastUpdatedTimestamp :: Lens' GetLoggerDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was last updated.

getLoggerDefinitionResponse_httpStatus :: Lens' GetLoggerDefinitionResponse Int

The response's http status code.