Network.AWS.Greengrass.GetLoggerDefinitionVersion

Description

Retrieves information about a logger definition version.

Creating a Request

data GetLoggerDefinitionVersion

See: newGetLoggerDefinitionVersion smart constructor.

Constructors

GetLoggerDefinitionVersion'

Fields

nextToken :: Maybe Text

The token for the next set of results, or ''null'' if there are no additional results.

loggerDefinitionVersionId :: Text

The ID of the logger definition version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListLoggerDefinitionVersions'' requests. If the version is the last one that was associated with a logger definition, the value also maps to the ''LatestVersion'' property of the corresponding ''DefinitionInformation'' object.

loggerDefinitionId :: Text

The ID of the logger definition.

newGetLoggerDefinitionVersion

Arguments

:: Text
$sel:loggerDefinitionVersionId:GetLoggerDefinitionVersion'

-> Text
$sel:loggerDefinitionId:GetLoggerDefinitionVersion'

-> GetLoggerDefinitionVersion

Create a value of GetLoggerDefinitionVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetLoggerDefinitionVersion', getLoggerDefinitionVersion_nextToken - The token for the next set of results, or ''null'' if there are no additional results.

$sel:loggerDefinitionVersionId:GetLoggerDefinitionVersion', getLoggerDefinitionVersion_loggerDefinitionVersionId - The ID of the logger definition version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListLoggerDefinitionVersions'' requests. If the version is the last one that was associated with a logger definition, the value also maps to the ''LatestVersion'' property of the corresponding ''DefinitionInformation'' object.

$sel:loggerDefinitionId:GetLoggerDefinitionVersion', getLoggerDefinitionVersion_loggerDefinitionId - The ID of the logger definition.

Request Lenses

getLoggerDefinitionVersion_nextToken :: Lens' GetLoggerDefinitionVersion (Maybe Text)

The token for the next set of results, or ''null'' if there are no additional results.

getLoggerDefinitionVersion_loggerDefinitionVersionId :: Lens' GetLoggerDefinitionVersion Text

The ID of the logger definition version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListLoggerDefinitionVersions'' requests. If the version is the last one that was associated with a logger definition, the value also maps to the ''LatestVersion'' property of the corresponding ''DefinitionInformation'' object.

getLoggerDefinitionVersion_loggerDefinitionId :: Lens' GetLoggerDefinitionVersion Text

The ID of the logger definition.

Destructuring the Response

data GetLoggerDefinitionVersionResponse

See: newGetLoggerDefinitionVersionResponse smart constructor.

Constructors

GetLoggerDefinitionVersionResponse'

Fields

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the logger definition version was created.

arn :: Maybe Text

The ARN of the logger definition version.

id :: Maybe Text

The ID of the logger definition version.

version :: Maybe Text

The version of the logger definition version.

definition :: Maybe LoggerDefinitionVersion

Information about the logger definition version.

httpStatus :: Int

The response's http status code.

newGetLoggerDefinitionVersionResponse

Arguments

:: Int
$sel:httpStatus:GetLoggerDefinitionVersionResponse'

-> GetLoggerDefinitionVersionResponse

Create a value of GetLoggerDefinitionVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTimestamp:GetLoggerDefinitionVersionResponse', getLoggerDefinitionVersionResponse_creationTimestamp - The time, in milliseconds since the epoch, when the logger definition version was created.

$sel:arn:GetLoggerDefinitionVersionResponse', getLoggerDefinitionVersionResponse_arn - The ARN of the logger definition version.

$sel:id:GetLoggerDefinitionVersionResponse', getLoggerDefinitionVersionResponse_id - The ID of the logger definition version.

$sel:version:GetLoggerDefinitionVersionResponse', getLoggerDefinitionVersionResponse_version - The version of the logger definition version.

$sel:definition:GetLoggerDefinitionVersionResponse', getLoggerDefinitionVersionResponse_definition - Information about the logger definition version.

$sel:httpStatus:GetLoggerDefinitionVersionResponse', getLoggerDefinitionVersionResponse_httpStatus - The response's http status code.

Response Lenses

getLoggerDefinitionVersionResponse_creationTimestamp :: Lens' GetLoggerDefinitionVersionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the logger definition version was created.

getLoggerDefinitionVersionResponse_arn :: Lens' GetLoggerDefinitionVersionResponse (Maybe Text)

The ARN of the logger definition version.

getLoggerDefinitionVersionResponse_id :: Lens' GetLoggerDefinitionVersionResponse (Maybe Text)

The ID of the logger definition version.

getLoggerDefinitionVersionResponse_version :: Lens' GetLoggerDefinitionVersionResponse (Maybe Text)

The version of the logger definition version.

getLoggerDefinitionVersionResponse_definition :: Lens' GetLoggerDefinitionVersionResponse (Maybe LoggerDefinitionVersion)

Information about the logger definition version.

getLoggerDefinitionVersionResponse_httpStatus :: Lens' GetLoggerDefinitionVersionResponse Int

The response's http status code.