Network.AWS.Greengrass.GetResourceDefinition

Description

Retrieves information about a resource definition, including its creation time and latest version.

Creating a Request

data GetResourceDefinition

See: newGetResourceDefinition smart constructor.

Constructors

GetResourceDefinition'

Fields

resourceDefinitionId :: Text

The ID of the resource definition.

newGetResourceDefinition

Arguments

:: Text
$sel:resourceDefinitionId:GetResourceDefinition'

-> GetResourceDefinition

Create a value of GetResourceDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceDefinitionId:GetResourceDefinition', getResourceDefinition_resourceDefinitionId - The ID of the resource definition.

Request Lenses

getResourceDefinition_resourceDefinitionId :: Lens' GetResourceDefinition Text

The ID of the resource definition.

Destructuring the Response

data GetResourceDefinitionResponse

See: newGetResourceDefinitionResponse smart constructor.

Constructors

GetResourceDefinitionResponse'

Fields

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was created.

latestVersionArn :: Maybe Text

The ARN of the latest version associated with the definition.

latestVersion :: Maybe Text

The ID of the latest version associated with the definition.

arn :: Maybe Text

The ARN of the definition.

id :: Maybe Text

The ID of the definition.

name :: Maybe Text

The name of the definition.

tags :: Maybe (HashMap Text Text)

Tag(s) attached to the resource arn.

lastUpdatedTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was last updated.

httpStatus :: Int

The response's http status code.

newGetResourceDefinitionResponse

Arguments

:: Int
$sel:httpStatus:GetResourceDefinitionResponse'

-> GetResourceDefinitionResponse

Create a value of GetResourceDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTimestamp:GetResourceDefinitionResponse', getResourceDefinitionResponse_creationTimestamp - The time, in milliseconds since the epoch, when the definition was created.

$sel:latestVersionArn:GetResourceDefinitionResponse', getResourceDefinitionResponse_latestVersionArn - The ARN of the latest version associated with the definition.

$sel:latestVersion:GetResourceDefinitionResponse', getResourceDefinitionResponse_latestVersion - The ID of the latest version associated with the definition.

$sel:arn:GetResourceDefinitionResponse', getResourceDefinitionResponse_arn - The ARN of the definition.

$sel:id:GetResourceDefinitionResponse', getResourceDefinitionResponse_id - The ID of the definition.

$sel:name:GetResourceDefinitionResponse', getResourceDefinitionResponse_name - The name of the definition.

$sel:tags:GetResourceDefinitionResponse', getResourceDefinitionResponse_tags - Tag(s) attached to the resource arn.

$sel:lastUpdatedTimestamp:GetResourceDefinitionResponse', getResourceDefinitionResponse_lastUpdatedTimestamp - The time, in milliseconds since the epoch, when the definition was last updated.

$sel:httpStatus:GetResourceDefinitionResponse', getResourceDefinitionResponse_httpStatus - The response's http status code.

Response Lenses

getResourceDefinitionResponse_creationTimestamp :: Lens' GetResourceDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was created.

getResourceDefinitionResponse_latestVersionArn :: Lens' GetResourceDefinitionResponse (Maybe Text)

The ARN of the latest version associated with the definition.

getResourceDefinitionResponse_latestVersion :: Lens' GetResourceDefinitionResponse (Maybe Text)

The ID of the latest version associated with the definition.

getResourceDefinitionResponse_arn :: Lens' GetResourceDefinitionResponse (Maybe Text)

The ARN of the definition.

getResourceDefinitionResponse_id :: Lens' GetResourceDefinitionResponse (Maybe Text)

The ID of the definition.

getResourceDefinitionResponse_name :: Lens' GetResourceDefinitionResponse (Maybe Text)

The name of the definition.

getResourceDefinitionResponse_tags :: Lens' GetResourceDefinitionResponse (Maybe (HashMap Text Text))

Tag(s) attached to the resource arn.

getResourceDefinitionResponse_lastUpdatedTimestamp :: Lens' GetResourceDefinitionResponse (Maybe Text)

The time, in milliseconds since the epoch, when the definition was last updated.

getResourceDefinitionResponse_httpStatus :: Lens' GetResourceDefinitionResponse Int

The response's http status code.