Network.AWS.Greengrass.GetServiceRoleForAccount

Description

Retrieves the service role that is attached to your account.

Creating a Request

data GetServiceRoleForAccount

See: newGetServiceRoleForAccount smart constructor.

Constructors

GetServiceRoleForAccount'

newGetServiceRoleForAccount :: GetServiceRoleForAccount

Create a value of GetServiceRoleForAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data GetServiceRoleForAccountResponse

See: newGetServiceRoleForAccountResponse smart constructor.

Constructors

GetServiceRoleForAccountResponse'

Fields

roleArn :: Maybe Text

The ARN of the role which is associated with the account.

associatedAt :: Maybe Text

The time when the service role was associated with the account.

httpStatus :: Int

The response's http status code.

newGetServiceRoleForAccountResponse

Arguments

:: Int
$sel:httpStatus:GetServiceRoleForAccountResponse'

-> GetServiceRoleForAccountResponse

Create a value of GetServiceRoleForAccountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleArn:GetServiceRoleForAccountResponse', getServiceRoleForAccountResponse_roleArn - The ARN of the role which is associated with the account.

$sel:associatedAt:GetServiceRoleForAccountResponse', getServiceRoleForAccountResponse_associatedAt - The time when the service role was associated with the account.

$sel:httpStatus:GetServiceRoleForAccountResponse', getServiceRoleForAccountResponse_httpStatus - The response's http status code.

Response Lenses

getServiceRoleForAccountResponse_roleArn :: Lens' GetServiceRoleForAccountResponse (Maybe Text)

The ARN of the role which is associated with the account.

getServiceRoleForAccountResponse_associatedAt :: Lens' GetServiceRoleForAccountResponse (Maybe Text)

The time when the service role was associated with the account.

getServiceRoleForAccountResponse_httpStatus :: Lens' GetServiceRoleForAccountResponse Int

The response's http status code.