Network.AWS.Greengrass.ListBulkDeploymentDetailedReports

Description

Gets a paginated list of the deployments that have been started in a bulk deployment operation, and their current deployment status.

This operation returns paginated results.

Creating a Request

data ListBulkDeploymentDetailedReports

See: newListBulkDeploymentDetailedReports smart constructor.

Constructors

ListBulkDeploymentDetailedReports'

Fields

nextToken :: Maybe Text

The token for the next set of results, or ''null'' if there are no additional results.

maxResults :: Maybe Text

The maximum number of results to be returned per request.

bulkDeploymentId :: Text

The ID of the bulk deployment.

newListBulkDeploymentDetailedReports

Arguments

:: Text
$sel:bulkDeploymentId:ListBulkDeploymentDetailedReports'

-> ListBulkDeploymentDetailedReports

Create a value of ListBulkDeploymentDetailedReports with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListBulkDeploymentDetailedReports', listBulkDeploymentDetailedReports_nextToken - The token for the next set of results, or ''null'' if there are no additional results.

$sel:maxResults:ListBulkDeploymentDetailedReports', listBulkDeploymentDetailedReports_maxResults - The maximum number of results to be returned per request.

$sel:bulkDeploymentId:ListBulkDeploymentDetailedReports', listBulkDeploymentDetailedReports_bulkDeploymentId - The ID of the bulk deployment.

Request Lenses

listBulkDeploymentDetailedReports_nextToken :: Lens' ListBulkDeploymentDetailedReports (Maybe Text)

The token for the next set of results, or ''null'' if there are no additional results.

listBulkDeploymentDetailedReports_maxResults :: Lens' ListBulkDeploymentDetailedReports (Maybe Text)

The maximum number of results to be returned per request.

listBulkDeploymentDetailedReports_bulkDeploymentId :: Lens' ListBulkDeploymentDetailedReports Text

The ID of the bulk deployment.

Destructuring the Response

data ListBulkDeploymentDetailedReportsResponse

See: newListBulkDeploymentDetailedReportsResponse smart constructor.

Constructors

ListBulkDeploymentDetailedReportsResponse'

Fields

nextToken :: Maybe Text

The token for the next set of results, or ''null'' if there are no additional results.

deployments :: Maybe [BulkDeploymentResult]

A list of the individual group deployments in the bulk deployment operation.

httpStatus :: Int

The response's http status code.

newListBulkDeploymentDetailedReportsResponse

Arguments

:: Int
$sel:httpStatus:ListBulkDeploymentDetailedReportsResponse'

-> ListBulkDeploymentDetailedReportsResponse

Create a value of ListBulkDeploymentDetailedReportsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListBulkDeploymentDetailedReports', listBulkDeploymentDetailedReportsResponse_nextToken - The token for the next set of results, or ''null'' if there are no additional results.

$sel:deployments:ListBulkDeploymentDetailedReportsResponse', listBulkDeploymentDetailedReportsResponse_deployments - A list of the individual group deployments in the bulk deployment operation.

$sel:httpStatus:ListBulkDeploymentDetailedReportsResponse', listBulkDeploymentDetailedReportsResponse_httpStatus - The response's http status code.

Response Lenses

listBulkDeploymentDetailedReportsResponse_nextToken :: Lens' ListBulkDeploymentDetailedReportsResponse (Maybe Text)

The token for the next set of results, or ''null'' if there are no additional results.

listBulkDeploymentDetailedReportsResponse_deployments :: Lens' ListBulkDeploymentDetailedReportsResponse (Maybe [BulkDeploymentResult])

A list of the individual group deployments in the bulk deployment operation.

listBulkDeploymentDetailedReportsResponse_httpStatus :: Lens' ListBulkDeploymentDetailedReportsResponse Int

The response's http status code.