Network.AWS.Greengrass.ListLoggerDefinitions

Description

Retrieves a list of logger definitions.

This operation returns paginated results.

Creating a Request

data ListLoggerDefinitions

See: newListLoggerDefinitions smart constructor.

Constructors

ListLoggerDefinitions'

Fields

nextToken :: Maybe Text

The token for the next set of results, or ''null'' if there are no additional results.

maxResults :: Maybe Text

The maximum number of results to be returned per request.

newListLoggerDefinitions :: ListLoggerDefinitions

Create a value of ListLoggerDefinitions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListLoggerDefinitions', listLoggerDefinitions_nextToken - The token for the next set of results, or ''null'' if there are no additional results.

$sel:maxResults:ListLoggerDefinitions', listLoggerDefinitions_maxResults - The maximum number of results to be returned per request.

Request Lenses

listLoggerDefinitions_nextToken :: Lens' ListLoggerDefinitions (Maybe Text)

The token for the next set of results, or ''null'' if there are no additional results.

listLoggerDefinitions_maxResults :: Lens' ListLoggerDefinitions (Maybe Text)

The maximum number of results to be returned per request.

Destructuring the Response

data ListLoggerDefinitionsResponse

See: newListLoggerDefinitionsResponse smart constructor.

Constructors

ListLoggerDefinitionsResponse'

Fields

nextToken :: Maybe Text

The token for the next set of results, or ''null'' if there are no additional results.

definitions :: Maybe [DefinitionInformation]

Information about a definition.

httpStatus :: Int

The response's http status code.

newListLoggerDefinitionsResponse

Arguments

:: Int
$sel:httpStatus:ListLoggerDefinitionsResponse'

-> ListLoggerDefinitionsResponse

Create a value of ListLoggerDefinitionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListLoggerDefinitions', listLoggerDefinitionsResponse_nextToken - The token for the next set of results, or ''null'' if there are no additional results.

$sel:definitions:ListLoggerDefinitionsResponse', listLoggerDefinitionsResponse_definitions - Information about a definition.

$sel:httpStatus:ListLoggerDefinitionsResponse', listLoggerDefinitionsResponse_httpStatus - The response's http status code.

Response Lenses

listLoggerDefinitionsResponse_nextToken :: Lens' ListLoggerDefinitionsResponse (Maybe Text)

The token for the next set of results, or ''null'' if there are no additional results.

listLoggerDefinitionsResponse_definitions :: Lens' ListLoggerDefinitionsResponse (Maybe [DefinitionInformation])

Information about a definition.

listLoggerDefinitionsResponse_httpStatus :: Lens' ListLoggerDefinitionsResponse Int

The response's http status code.