Network.AWS.Greengrass.ListResourceDefinitionVersions

Description

Lists the versions of a resource definition.

This operation returns paginated results.

Creating a Request

data ListResourceDefinitionVersions

See: newListResourceDefinitionVersions smart constructor.

Constructors

ListResourceDefinitionVersions'

Fields

nextToken :: Maybe Text

The token for the next set of results, or ''null'' if there are no additional results.

maxResults :: Maybe Text

The maximum number of results to be returned per request.

resourceDefinitionId :: Text

The ID of the resource definition.

newListResourceDefinitionVersions

Arguments

:: Text
$sel:resourceDefinitionId:ListResourceDefinitionVersions'

-> ListResourceDefinitionVersions

Create a value of ListResourceDefinitionVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListResourceDefinitionVersions', listResourceDefinitionVersions_nextToken - The token for the next set of results, or ''null'' if there are no additional results.

$sel:maxResults:ListResourceDefinitionVersions', listResourceDefinitionVersions_maxResults - The maximum number of results to be returned per request.

$sel:resourceDefinitionId:ListResourceDefinitionVersions', listResourceDefinitionVersions_resourceDefinitionId - The ID of the resource definition.

Request Lenses

listResourceDefinitionVersions_nextToken :: Lens' ListResourceDefinitionVersions (Maybe Text)

The token for the next set of results, or ''null'' if there are no additional results.

listResourceDefinitionVersions_maxResults :: Lens' ListResourceDefinitionVersions (Maybe Text)

The maximum number of results to be returned per request.

listResourceDefinitionVersions_resourceDefinitionId :: Lens' ListResourceDefinitionVersions Text

The ID of the resource definition.

Destructuring the Response

data ListResourceDefinitionVersionsResponse

See: newListResourceDefinitionVersionsResponse smart constructor.

Constructors

ListResourceDefinitionVersionsResponse'

Fields

nextToken :: Maybe Text

The token for the next set of results, or ''null'' if there are no additional results.

versions :: Maybe [VersionInformation]

Information about a version.

httpStatus :: Int

The response's http status code.

newListResourceDefinitionVersionsResponse

Arguments

:: Int
$sel:httpStatus:ListResourceDefinitionVersionsResponse'

-> ListResourceDefinitionVersionsResponse

Create a value of ListResourceDefinitionVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListResourceDefinitionVersions', listResourceDefinitionVersionsResponse_nextToken - The token for the next set of results, or ''null'' if there are no additional results.

$sel:versions:ListResourceDefinitionVersionsResponse', listResourceDefinitionVersionsResponse_versions - Information about a version.

$sel:httpStatus:ListResourceDefinitionVersionsResponse', listResourceDefinitionVersionsResponse_httpStatus - The response's http status code.

Response Lenses

listResourceDefinitionVersionsResponse_nextToken :: Lens' ListResourceDefinitionVersionsResponse (Maybe Text)

The token for the next set of results, or ''null'' if there are no additional results.

listResourceDefinitionVersionsResponse_versions :: Lens' ListResourceDefinitionVersionsResponse (Maybe [VersionInformation])

Information about a version.

listResourceDefinitionVersionsResponse_httpStatus :: Lens' ListResourceDefinitionVersionsResponse Int

The response's http status code.