Network.AWS.Greengrass.ListSubscriptionDefinitions

Description

Retrieves a list of subscription definitions.

This operation returns paginated results.

Creating a Request

data ListSubscriptionDefinitions

See: newListSubscriptionDefinitions smart constructor.

Constructors

ListSubscriptionDefinitions'

Fields

nextToken :: Maybe Text

The token for the next set of results, or ''null'' if there are no additional results.

maxResults :: Maybe Text

The maximum number of results to be returned per request.

newListSubscriptionDefinitions :: ListSubscriptionDefinitions

Create a value of ListSubscriptionDefinitions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSubscriptionDefinitions', listSubscriptionDefinitions_nextToken - The token for the next set of results, or ''null'' if there are no additional results.

$sel:maxResults:ListSubscriptionDefinitions', listSubscriptionDefinitions_maxResults - The maximum number of results to be returned per request.

Request Lenses

listSubscriptionDefinitions_nextToken :: Lens' ListSubscriptionDefinitions (Maybe Text)

The token for the next set of results, or ''null'' if there are no additional results.

listSubscriptionDefinitions_maxResults :: Lens' ListSubscriptionDefinitions (Maybe Text)

The maximum number of results to be returned per request.

Destructuring the Response

data ListSubscriptionDefinitionsResponse

See: newListSubscriptionDefinitionsResponse smart constructor.

Constructors

ListSubscriptionDefinitionsResponse'

Fields

nextToken :: Maybe Text

The token for the next set of results, or ''null'' if there are no additional results.

definitions :: Maybe [DefinitionInformation]

Information about a definition.

httpStatus :: Int

The response's http status code.

newListSubscriptionDefinitionsResponse

Arguments

:: Int
$sel:httpStatus:ListSubscriptionDefinitionsResponse'

-> ListSubscriptionDefinitionsResponse

Create a value of ListSubscriptionDefinitionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListSubscriptionDefinitions', listSubscriptionDefinitionsResponse_nextToken - The token for the next set of results, or ''null'' if there are no additional results.

$sel:definitions:ListSubscriptionDefinitionsResponse', listSubscriptionDefinitionsResponse_definitions - Information about a definition.

$sel:httpStatus:ListSubscriptionDefinitionsResponse', listSubscriptionDefinitionsResponse_httpStatus - The response's http status code.

Response Lenses

listSubscriptionDefinitionsResponse_nextToken :: Lens' ListSubscriptionDefinitionsResponse (Maybe Text)

The token for the next set of results, or ''null'' if there are no additional results.

listSubscriptionDefinitionsResponse_definitions :: Lens' ListSubscriptionDefinitionsResponse (Maybe [DefinitionInformation])

Information about a definition.

listSubscriptionDefinitionsResponse_httpStatus :: Lens' ListSubscriptionDefinitionsResponse Int

The response's http status code.