Network.AWS.Greengrass.ResetDeployments

Description

Resets a group's deployments.

Creating a Request

data ResetDeployments

Information needed to reset deployments.

See: newResetDeployments smart constructor.

Constructors

ResetDeployments'

Fields

force :: Maybe Bool

If true, performs a best-effort only core reset.

amznClientToken :: Maybe Text

A client token used to correlate requests and responses.

groupId :: Text

The ID of the Greengrass group.

newResetDeployments

Arguments

:: Text
$sel:groupId:ResetDeployments'

-> ResetDeployments

Create a value of ResetDeployments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:force:ResetDeployments', resetDeployments_force - If true, performs a best-effort only core reset.

$sel:amznClientToken:ResetDeployments', resetDeployments_amznClientToken - A client token used to correlate requests and responses.

$sel:groupId:ResetDeployments', resetDeployments_groupId - The ID of the Greengrass group.

Request Lenses

resetDeployments_force :: Lens' ResetDeployments (Maybe Bool)

If true, performs a best-effort only core reset.

resetDeployments_amznClientToken :: Lens' ResetDeployments (Maybe Text)

A client token used to correlate requests and responses.

resetDeployments_groupId :: Lens' ResetDeployments Text

The ID of the Greengrass group.

Destructuring the Response

data ResetDeploymentsResponse

See: newResetDeploymentsResponse smart constructor.

Constructors

ResetDeploymentsResponse'

Fields

deploymentId :: Maybe Text

The ID of the deployment.

deploymentArn :: Maybe Text

The ARN of the deployment.

httpStatus :: Int

The response's http status code.

newResetDeploymentsResponse

Arguments

:: Int
$sel:httpStatus:ResetDeploymentsResponse'

-> ResetDeploymentsResponse

Create a value of ResetDeploymentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentId:ResetDeploymentsResponse', resetDeploymentsResponse_deploymentId - The ID of the deployment.

$sel:deploymentArn:ResetDeploymentsResponse', resetDeploymentsResponse_deploymentArn - The ARN of the deployment.

$sel:httpStatus:ResetDeploymentsResponse', resetDeploymentsResponse_httpStatus - The response's http status code.

Response Lenses

resetDeploymentsResponse_deploymentId :: Lens' ResetDeploymentsResponse (Maybe Text)

The ID of the deployment.

resetDeploymentsResponse_deploymentArn :: Lens' ResetDeploymentsResponse (Maybe Text)

The ARN of the deployment.

resetDeploymentsResponse_httpStatus :: Lens' ResetDeploymentsResponse Int

The response's http status code.