Network.AWS.Greengrass.Types.BulkDeploymentMetrics

Description

Documentation

data BulkDeploymentMetrics

Relevant metrics on input records processed during bulk deployment.

See: newBulkDeploymentMetrics smart constructor.

Constructors

BulkDeploymentMetrics'

Fields

recordsProcessed :: Maybe Int

The total number of group records from the input file that have been processed so far, or attempted.

retryAttempts :: Maybe Int

The total number of deployment attempts that returned a retryable error. For example, a retry is triggered if the attempt to deploy a group returns a throttling error. ''StartBulkDeployment'' retries a group deployment up to five times.

invalidInputRecords :: Maybe Int

The total number of records that returned a non-retryable error. For example, this can occur if a group record from the input file uses an invalid format or specifies a nonexistent group version, or if the execution role doesn't grant permission to deploy a group or group version.

newBulkDeploymentMetrics :: BulkDeploymentMetrics

Create a value of BulkDeploymentMetrics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recordsProcessed:BulkDeploymentMetrics', bulkDeploymentMetrics_recordsProcessed - The total number of group records from the input file that have been processed so far, or attempted.

$sel:retryAttempts:BulkDeploymentMetrics', bulkDeploymentMetrics_retryAttempts - The total number of deployment attempts that returned a retryable error. For example, a retry is triggered if the attempt to deploy a group returns a throttling error. ''StartBulkDeployment'' retries a group deployment up to five times.

$sel:invalidInputRecords:BulkDeploymentMetrics', bulkDeploymentMetrics_invalidInputRecords - The total number of records that returned a non-retryable error. For example, this can occur if a group record from the input file uses an invalid format or specifies a nonexistent group version, or if the execution role doesn't grant permission to deploy a group or group version.

bulkDeploymentMetrics_recordsProcessed :: Lens' BulkDeploymentMetrics (Maybe Int)

The total number of group records from the input file that have been processed so far, or attempted.

bulkDeploymentMetrics_retryAttempts :: Lens' BulkDeploymentMetrics (Maybe Int)

The total number of deployment attempts that returned a retryable error. For example, a retry is triggered if the attempt to deploy a group returns a throttling error. ''StartBulkDeployment'' retries a group deployment up to five times.

bulkDeploymentMetrics_invalidInputRecords :: Lens' BulkDeploymentMetrics (Maybe Int)

The total number of records that returned a non-retryable error. For example, this can occur if a group record from the input file uses an invalid format or specifies a nonexistent group version, or if the execution role doesn't grant permission to deploy a group or group version.