Network.AWS.Greengrass.Types.DefinitionInformation

Description

Documentation

data DefinitionInformation

Information about a definition.

See: newDefinitionInformation smart constructor.

Constructors

DefinitionInformation'

Fields

creationTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was created.

latestVersionArn :: Maybe Text

The ARN of the latest version associated with the definition.

latestVersion :: Maybe Text

The ID of the latest version associated with the definition.

arn :: Maybe Text

The ARN of the definition.

id :: Maybe Text

The ID of the definition.

name :: Maybe Text

The name of the definition.

tags :: Maybe (HashMap Text Text)

Tag(s) attached to the resource arn.

lastUpdatedTimestamp :: Maybe Text

The time, in milliseconds since the epoch, when the definition was last updated.

newDefinitionInformation :: DefinitionInformation

Create a value of DefinitionInformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTimestamp:DefinitionInformation', definitionInformation_creationTimestamp - The time, in milliseconds since the epoch, when the definition was created.

$sel:latestVersionArn:DefinitionInformation', definitionInformation_latestVersionArn - The ARN of the latest version associated with the definition.

$sel:latestVersion:DefinitionInformation', definitionInformation_latestVersion - The ID of the latest version associated with the definition.

$sel:arn:DefinitionInformation', definitionInformation_arn - The ARN of the definition.

$sel:id:DefinitionInformation', definitionInformation_id - The ID of the definition.

$sel:name:DefinitionInformation', definitionInformation_name - The name of the definition.

$sel:tags:DefinitionInformation', definitionInformation_tags - Tag(s) attached to the resource arn.

$sel:lastUpdatedTimestamp:DefinitionInformation', definitionInformation_lastUpdatedTimestamp - The time, in milliseconds since the epoch, when the definition was last updated.

definitionInformation_creationTimestamp :: Lens' DefinitionInformation (Maybe Text)

The time, in milliseconds since the epoch, when the definition was created.

definitionInformation_latestVersionArn :: Lens' DefinitionInformation (Maybe Text)

The ARN of the latest version associated with the definition.

definitionInformation_latestVersion :: Lens' DefinitionInformation (Maybe Text)

The ID of the latest version associated with the definition.

definitionInformation_arn :: Lens' DefinitionInformation (Maybe Text)

The ARN of the definition.

definitionInformation_id :: Lens' DefinitionInformation (Maybe Text)

The ID of the definition.

definitionInformation_name :: Lens' DefinitionInformation (Maybe Text)

The name of the definition.

definitionInformation_tags :: Lens' DefinitionInformation (Maybe (HashMap Text Text))

Tag(s) attached to the resource arn.

definitionInformation_lastUpdatedTimestamp :: Lens' DefinitionInformation (Maybe Text)

The time, in milliseconds since the epoch, when the definition was last updated.