Network.AWS.Greengrass.Types.Deployment

Description

Documentation

data Deployment

Information about a deployment.

See: newDeployment smart constructor.

Constructors

Deployment'

Fields

deploymentId :: Maybe Text

The ID of the deployment.

deploymentArn :: Maybe Text

The ARN of the deployment.

createdAt :: Maybe Text

The time, in milliseconds since the epoch, when the deployment was created.

deploymentType :: Maybe DeploymentType

The type of the deployment.

groupArn :: Maybe Text

The ARN of the group for this deployment.

newDeployment :: Deployment

Create a value of Deployment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentId:Deployment', deployment_deploymentId - The ID of the deployment.

$sel:deploymentArn:Deployment', deployment_deploymentArn - The ARN of the deployment.

$sel:createdAt:Deployment', deployment_createdAt - The time, in milliseconds since the epoch, when the deployment was created.

$sel:deploymentType:Deployment', deployment_deploymentType - The type of the deployment.

$sel:groupArn:Deployment', deployment_groupArn - The ARN of the group for this deployment.

deployment_deploymentId :: Lens' Deployment (Maybe Text)

The ID of the deployment.

deployment_deploymentArn :: Lens' Deployment (Maybe Text)

The ARN of the deployment.

deployment_createdAt :: Lens' Deployment (Maybe Text)

The time, in milliseconds since the epoch, when the deployment was created.

deployment_deploymentType :: Lens' Deployment (Maybe DeploymentType)

The type of the deployment.

deployment_groupArn :: Lens' Deployment (Maybe Text)

The ARN of the group for this deployment.