Network.AWS.Greengrass.Types.Function

Description

Documentation

data Function

Information about a Lambda function.

See: newFunction smart constructor.

Constructors

Function'

Fields

functionArn :: Maybe Text

The ARN of the Lambda function.

functionConfiguration :: Maybe FunctionConfiguration

The configuration of the Lambda function.

id :: Text

A descriptive or arbitrary ID for the function. This value must be unique within the function definition version. Max length is 128 characters with pattern ''[a-zA-Z0-9:_-]+''.

newFunction

Arguments

:: Text
$sel:id:Function'

-> Function

Create a value of Function with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:functionArn:Function', function_functionArn - The ARN of the Lambda function.

$sel:functionConfiguration:Function', function_functionConfiguration - The configuration of the Lambda function.

$sel:id:Function', function_id - A descriptive or arbitrary ID for the function. This value must be unique within the function definition version. Max length is 128 characters with pattern ''[a-zA-Z0-9:_-]+''.

function_functionArn :: Lens' Function (Maybe Text)

The ARN of the Lambda function.

function_functionConfiguration :: Lens' Function (Maybe FunctionConfiguration)

The configuration of the Lambda function.

function_id :: Lens' Function Text

A descriptive or arbitrary ID for the function. This value must be unique within the function definition version. Max length is 128 characters with pattern ''[a-zA-Z0-9:_-]+''.