Network.AWS.Greengrass.Types.FunctionRunAsConfig

Description

Documentation

data FunctionRunAsConfig

Specifies the user and group whose permissions are used when running the Lambda function. You can specify one or both values to override the default values. We recommend that you avoid running as root unless absolutely necessary to minimize the risk of unintended changes or malicious attacks. To run as root, you must set ''IsolationMode'' to ''NoContainer'' and update config.json in ''greengrass-root/config'' to set ''allowFunctionsToRunAsRoot'' to ''yes''.

See: newFunctionRunAsConfig smart constructor.

Constructors

FunctionRunAsConfig'

Fields

gid :: Maybe Int

The group ID whose permissions are used to run a Lambda function.

uid :: Maybe Int

The user ID whose permissions are used to run a Lambda function.

newFunctionRunAsConfig :: FunctionRunAsConfig

Create a value of FunctionRunAsConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gid:FunctionRunAsConfig', functionRunAsConfig_gid - The group ID whose permissions are used to run a Lambda function.

$sel:uid:FunctionRunAsConfig', functionRunAsConfig_uid - The user ID whose permissions are used to run a Lambda function.

functionRunAsConfig_gid :: Lens' FunctionRunAsConfig (Maybe Int)

The group ID whose permissions are used to run a Lambda function.

functionRunAsConfig_uid :: Lens' FunctionRunAsConfig (Maybe Int)

The user ID whose permissions are used to run a Lambda function.