Network.AWS.Greengrass.Types.GroupVersion

Description

Documentation

data GroupVersion

Information about a group version.

See: newGroupVersion smart constructor.

Constructors

GroupVersion'

Fields

subscriptionDefinitionVersionArn :: Maybe Text

The ARN of the subscription definition version for this group.

coreDefinitionVersionArn :: Maybe Text

The ARN of the core definition version for this group.

connectorDefinitionVersionArn :: Maybe Text

The ARN of the connector definition version for this group.

loggerDefinitionVersionArn :: Maybe Text

The ARN of the logger definition version for this group.

resourceDefinitionVersionArn :: Maybe Text

The ARN of the resource definition version for this group.

functionDefinitionVersionArn :: Maybe Text

The ARN of the function definition version for this group.

deviceDefinitionVersionArn :: Maybe Text

The ARN of the device definition version for this group.

newGroupVersion :: GroupVersion

Create a value of GroupVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subscriptionDefinitionVersionArn:GroupVersion', groupVersion_subscriptionDefinitionVersionArn - The ARN of the subscription definition version for this group.

$sel:coreDefinitionVersionArn:GroupVersion', groupVersion_coreDefinitionVersionArn - The ARN of the core definition version for this group.

$sel:connectorDefinitionVersionArn:GroupVersion', groupVersion_connectorDefinitionVersionArn - The ARN of the connector definition version for this group.

$sel:loggerDefinitionVersionArn:GroupVersion', groupVersion_loggerDefinitionVersionArn - The ARN of the logger definition version for this group.

$sel:resourceDefinitionVersionArn:GroupVersion', groupVersion_resourceDefinitionVersionArn - The ARN of the resource definition version for this group.

$sel:functionDefinitionVersionArn:GroupVersion', groupVersion_functionDefinitionVersionArn - The ARN of the function definition version for this group.

$sel:deviceDefinitionVersionArn:GroupVersion', groupVersion_deviceDefinitionVersionArn - The ARN of the device definition version for this group.

groupVersion_subscriptionDefinitionVersionArn :: Lens' GroupVersion (Maybe Text)

The ARN of the subscription definition version for this group.

groupVersion_coreDefinitionVersionArn :: Lens' GroupVersion (Maybe Text)

The ARN of the core definition version for this group.

groupVersion_connectorDefinitionVersionArn :: Lens' GroupVersion (Maybe Text)

The ARN of the connector definition version for this group.

groupVersion_loggerDefinitionVersionArn :: Lens' GroupVersion (Maybe Text)

The ARN of the logger definition version for this group.

groupVersion_resourceDefinitionVersionArn :: Lens' GroupVersion (Maybe Text)

The ARN of the resource definition version for this group.

groupVersion_functionDefinitionVersionArn :: Lens' GroupVersion (Maybe Text)

The ARN of the function definition version for this group.

groupVersion_deviceDefinitionVersionArn :: Lens' GroupVersion (Maybe Text)

The ARN of the device definition version for this group.