Network.AWS.Greengrass.UpdateConnectivityInfo

Description

Updates the connectivity information for the core. Any devices that belong to the group which has this core will receive this information in order to find the location of the core and connect to it.

Creating a Request

data UpdateConnectivityInfo

Connectivity information.

See: newUpdateConnectivityInfo smart constructor.

Constructors

UpdateConnectivityInfo'

Fields

connectivityInfo :: Maybe [ConnectivityInfo]

A list of connectivity info.

thingName :: Text

The thing name.

newUpdateConnectivityInfo

Arguments

:: Text
$sel:thingName:UpdateConnectivityInfo'

-> UpdateConnectivityInfo

Create a value of UpdateConnectivityInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectivityInfo:UpdateConnectivityInfo', updateConnectivityInfo_connectivityInfo - A list of connectivity info.

$sel:thingName:UpdateConnectivityInfo', updateConnectivityInfo_thingName - The thing name.

Request Lenses

updateConnectivityInfo_connectivityInfo :: Lens' UpdateConnectivityInfo (Maybe [ConnectivityInfo])

A list of connectivity info.

updateConnectivityInfo_thingName :: Lens' UpdateConnectivityInfo Text

The thing name.

Destructuring the Response

data UpdateConnectivityInfoResponse

See: newUpdateConnectivityInfoResponse smart constructor.

Constructors

UpdateConnectivityInfoResponse'

Fields

message :: Maybe Text

A message about the connectivity info update request.

version :: Maybe Text

The new version of the connectivity info.

httpStatus :: Int

The response's http status code.

newUpdateConnectivityInfoResponse

Arguments

:: Int
$sel:httpStatus:UpdateConnectivityInfoResponse'

-> UpdateConnectivityInfoResponse

Create a value of UpdateConnectivityInfoResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:message:UpdateConnectivityInfoResponse', updateConnectivityInfoResponse_message - A message about the connectivity info update request.

$sel:version:UpdateConnectivityInfoResponse', updateConnectivityInfoResponse_version - The new version of the connectivity info.

$sel:httpStatus:UpdateConnectivityInfoResponse', updateConnectivityInfoResponse_httpStatus - The response's http status code.

Response Lenses

updateConnectivityInfoResponse_message :: Lens' UpdateConnectivityInfoResponse (Maybe Text)

A message about the connectivity info update request.

updateConnectivityInfoResponse_version :: Lens' UpdateConnectivityInfoResponse (Maybe Text)

The new version of the connectivity info.

updateConnectivityInfoResponse_httpStatus :: Lens' UpdateConnectivityInfoResponse Int

The response's http status code.