Network.AWS.GreengrassV2.GetDeployment

Description

Gets a deployment. Deployments define the components that run on Greengrass core devices.

Creating a Request

data GetDeployment

See: newGetDeployment smart constructor.

Constructors

GetDeployment'

Fields

deploymentId :: Text

The ID of the deployment.

newGetDeployment

Arguments

:: Text
$sel:deploymentId:GetDeployment'

-> GetDeployment

Create a value of GetDeployment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentId:GetDeployment', getDeployment_deploymentId - The ID of the deployment.

Request Lenses

getDeployment_deploymentId :: Lens' GetDeployment Text

The ID of the deployment.

Destructuring the Response

data GetDeploymentResponse

See: newGetDeploymentResponse smart constructor.

Constructors

GetDeploymentResponse'

Fields

targetArn :: Maybe Text

The ARN of the target IoT thing or thing group.

components :: Maybe (HashMap Text ComponentDeploymentSpecification)

The components to deploy. This is a dictionary, where each key is the name of a component, and each key's value is the version and configuration to deploy for that component.

deploymentId :: Maybe Text

The ID of the deployment.

iotJobId :: Maybe Text

The ID of the IoT job that applies the deployment to target devices.

iotJobArn :: Maybe Text

The ARN of the IoT job that applies the deployment to target devices.

deploymentPolicies :: Maybe DeploymentPolicies

The deployment policies for the deployment. These policies define how the deployment updates components and handles failure.

creationTimestamp :: Maybe POSIX

The time at which the deployment was created, expressed in ISO 8601 format.

iotJobConfiguration :: Maybe DeploymentIoTJobConfiguration

The job configuration for the deployment configuration. The job configuration specifies the rollout, timeout, and stop configurations for the deployment configuration.

deploymentStatus :: Maybe DeploymentStatus

The status of the deployment.

isLatestForTarget :: Maybe Bool

Whether or not the deployment is the latest revision for its target.

revisionId :: Maybe Text

The revision number of the deployment.

deploymentName :: Maybe Text

The name of the deployment.

tags :: Maybe (HashMap Text Text)

A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

httpStatus :: Int

The response's http status code.

newGetDeploymentResponse

Arguments

:: Int
$sel:httpStatus:GetDeploymentResponse'

-> GetDeploymentResponse

Create a value of GetDeploymentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetArn:GetDeploymentResponse', getDeploymentResponse_targetArn - The ARN of the target IoT thing or thing group.

$sel:components:GetDeploymentResponse', getDeploymentResponse_components - The components to deploy. This is a dictionary, where each key is the name of a component, and each key's value is the version and configuration to deploy for that component.

$sel:deploymentId:GetDeployment', getDeploymentResponse_deploymentId - The ID of the deployment.

$sel:iotJobId:GetDeploymentResponse', getDeploymentResponse_iotJobId - The ID of the IoT job that applies the deployment to target devices.

$sel:iotJobArn:GetDeploymentResponse', getDeploymentResponse_iotJobArn - The ARN of the IoT job that applies the deployment to target devices.

$sel:deploymentPolicies:GetDeploymentResponse', getDeploymentResponse_deploymentPolicies - The deployment policies for the deployment. These policies define how the deployment updates components and handles failure.

$sel:creationTimestamp:GetDeploymentResponse', getDeploymentResponse_creationTimestamp - The time at which the deployment was created, expressed in ISO 8601 format.

$sel:iotJobConfiguration:GetDeploymentResponse', getDeploymentResponse_iotJobConfiguration - The job configuration for the deployment configuration. The job configuration specifies the rollout, timeout, and stop configurations for the deployment configuration.

$sel:deploymentStatus:GetDeploymentResponse', getDeploymentResponse_deploymentStatus - The status of the deployment.

$sel:isLatestForTarget:GetDeploymentResponse', getDeploymentResponse_isLatestForTarget - Whether or not the deployment is the latest revision for its target.

$sel:revisionId:GetDeploymentResponse', getDeploymentResponse_revisionId - The revision number of the deployment.

$sel:deploymentName:GetDeploymentResponse', getDeploymentResponse_deploymentName - The name of the deployment.

$sel:tags:GetDeploymentResponse', getDeploymentResponse_tags - A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

$sel:httpStatus:GetDeploymentResponse', getDeploymentResponse_httpStatus - The response's http status code.

Response Lenses

getDeploymentResponse_targetArn :: Lens' GetDeploymentResponse (Maybe Text)

The ARN of the target IoT thing or thing group.

getDeploymentResponse_components :: Lens' GetDeploymentResponse (Maybe (HashMap Text ComponentDeploymentSpecification))

The components to deploy. This is a dictionary, where each key is the name of a component, and each key's value is the version and configuration to deploy for that component.

getDeploymentResponse_deploymentId :: Lens' GetDeploymentResponse (Maybe Text)

The ID of the deployment.

getDeploymentResponse_iotJobId :: Lens' GetDeploymentResponse (Maybe Text)

The ID of the IoT job that applies the deployment to target devices.

getDeploymentResponse_iotJobArn :: Lens' GetDeploymentResponse (Maybe Text)

The ARN of the IoT job that applies the deployment to target devices.

getDeploymentResponse_deploymentPolicies :: Lens' GetDeploymentResponse (Maybe DeploymentPolicies)

The deployment policies for the deployment. These policies define how the deployment updates components and handles failure.

getDeploymentResponse_creationTimestamp :: Lens' GetDeploymentResponse (Maybe UTCTime)

The time at which the deployment was created, expressed in ISO 8601 format.

getDeploymentResponse_iotJobConfiguration :: Lens' GetDeploymentResponse (Maybe DeploymentIoTJobConfiguration)

The job configuration for the deployment configuration. The job configuration specifies the rollout, timeout, and stop configurations for the deployment configuration.

getDeploymentResponse_deploymentStatus :: Lens' GetDeploymentResponse (Maybe DeploymentStatus)

The status of the deployment.

getDeploymentResponse_isLatestForTarget :: Lens' GetDeploymentResponse (Maybe Bool)

Whether or not the deployment is the latest revision for its target.

getDeploymentResponse_revisionId :: Lens' GetDeploymentResponse (Maybe Text)

The revision number of the deployment.

getDeploymentResponse_deploymentName :: Lens' GetDeploymentResponse (Maybe Text)

The name of the deployment.

getDeploymentResponse_tags :: Lens' GetDeploymentResponse (Maybe (HashMap Text Text))

A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

getDeploymentResponse_httpStatus :: Lens' GetDeploymentResponse Int

The response's http status code.