Network.AWS.GreengrassV2.Lens

Description

Operations

ListComponentVersions

listComponentVersions_nextToken :: Lens' ListComponentVersions (Maybe Text)

The token to be used for the next set of paginated results.

listComponentVersions_maxResults :: Lens' ListComponentVersions (Maybe Natural)

The maximum number of results to be returned per paginated request.

listComponentVersions_arn :: Lens' ListComponentVersions Text

The ARN of the component version.

listComponentVersionsResponse_nextToken :: Lens' ListComponentVersionsResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

listComponentVersionsResponse_componentVersions :: Lens' ListComponentVersionsResponse (Maybe [ComponentVersionListItem])

A list of versions that exist for the component.

listComponentVersionsResponse_httpStatus :: Lens' ListComponentVersionsResponse Int

The response's http status code.

GetDeployment

getDeployment_deploymentId :: Lens' GetDeployment Text

The ID of the deployment.

getDeploymentResponse_targetArn :: Lens' GetDeploymentResponse (Maybe Text)

The ARN of the target IoT thing or thing group.

getDeploymentResponse_components :: Lens' GetDeploymentResponse (Maybe (HashMap Text ComponentDeploymentSpecification))

The components to deploy. This is a dictionary, where each key is the name of a component, and each key's value is the version and configuration to deploy for that component.

getDeploymentResponse_deploymentId :: Lens' GetDeploymentResponse (Maybe Text)

The ID of the deployment.

getDeploymentResponse_iotJobId :: Lens' GetDeploymentResponse (Maybe Text)

The ID of the IoT job that applies the deployment to target devices.

getDeploymentResponse_iotJobArn :: Lens' GetDeploymentResponse (Maybe Text)

The ARN of the IoT job that applies the deployment to target devices.

getDeploymentResponse_deploymentPolicies :: Lens' GetDeploymentResponse (Maybe DeploymentPolicies)

The deployment policies for the deployment. These policies define how the deployment updates components and handles failure.

getDeploymentResponse_creationTimestamp :: Lens' GetDeploymentResponse (Maybe UTCTime)

The time at which the deployment was created, expressed in ISO 8601 format.

getDeploymentResponse_iotJobConfiguration :: Lens' GetDeploymentResponse (Maybe DeploymentIoTJobConfiguration)

The job configuration for the deployment configuration. The job configuration specifies the rollout, timeout, and stop configurations for the deployment configuration.

getDeploymentResponse_deploymentStatus :: Lens' GetDeploymentResponse (Maybe DeploymentStatus)

The status of the deployment.

getDeploymentResponse_isLatestForTarget :: Lens' GetDeploymentResponse (Maybe Bool)

Whether or not the deployment is the latest revision for its target.

getDeploymentResponse_revisionId :: Lens' GetDeploymentResponse (Maybe Text)

The revision number of the deployment.

getDeploymentResponse_deploymentName :: Lens' GetDeploymentResponse (Maybe Text)

The name of the deployment.

getDeploymentResponse_tags :: Lens' GetDeploymentResponse (Maybe (HashMap Text Text))

A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

getDeploymentResponse_httpStatus :: Lens' GetDeploymentResponse Int

The response's http status code.

DescribeComponent

describeComponent_arn :: Lens' DescribeComponent Text

The ARN of the component version.

describeComponentResponse_platforms :: Lens' DescribeComponentResponse (Maybe [ComponentPlatform])

The platforms that the component version supports.

describeComponentResponse_status :: Lens' DescribeComponentResponse (Maybe CloudComponentStatus)

The status of the component version in IoT Greengrass V2. This status is different from the status of the component on a core device.

describeComponentResponse_arn :: Lens' DescribeComponentResponse (Maybe Text)

The ARN of the component version.

describeComponentResponse_componentVersion :: Lens' DescribeComponentResponse (Maybe Text)

The version of the component.

describeComponentResponse_creationTimestamp :: Lens' DescribeComponentResponse (Maybe UTCTime)

The time at which the component was created, expressed in ISO 8601 format.

describeComponentResponse_componentName :: Lens' DescribeComponentResponse (Maybe Text)

The name of the component.

describeComponentResponse_publisher :: Lens' DescribeComponentResponse (Maybe Text)

The publisher of the component version.

describeComponentResponse_description :: Lens' DescribeComponentResponse (Maybe Text)

The description of the component version.

describeComponentResponse_tags :: Lens' DescribeComponentResponse (Maybe (HashMap Text Text))

A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

describeComponentResponse_httpStatus :: Lens' DescribeComponentResponse Int

The response's http status code.

DeleteCoreDevice

deleteCoreDevice_coreDeviceThingName :: Lens' DeleteCoreDevice Text

The name of the core device. This is also the name of the IoT thing.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The ARN of the resource.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

GetCoreDevice

getCoreDevice_coreDeviceThingName :: Lens' GetCoreDevice Text

The name of the core device. This is also the name of the IoT thing.

getCoreDeviceResponse_status :: Lens' GetCoreDeviceResponse (Maybe CoreDeviceStatus)

The status of the core device. The core device status can be:

HEALTHY – The IoT Greengrass Core software and all components run on the core device without issue.

UNHEALTHY – The IoT Greengrass Core software or a component is in a failed state on the core device.

getCoreDeviceResponse_platform :: Lens' GetCoreDeviceResponse (Maybe Text)

The operating system platform that the core device runs.

getCoreDeviceResponse_architecture :: Lens' GetCoreDeviceResponse (Maybe Text)

The computer architecture of the core device.

getCoreDeviceResponse_coreDeviceThingName :: Lens' GetCoreDeviceResponse (Maybe Text)

The name of the core device. This is also the name of the IoT thing.

getCoreDeviceResponse_tags :: Lens' GetCoreDeviceResponse (Maybe (HashMap Text Text))

A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

getCoreDeviceResponse_coreVersion :: Lens' GetCoreDeviceResponse (Maybe Text)

The version of the IoT Greengrass Core software that the core device runs. This version is equivalent to the version of the Greengrass nucleus component that runs on the core device. For more information, see the Greengrass nucleus component in the IoT Greengrass V2 Developer Guide.

getCoreDeviceResponse_lastStatusUpdateTimestamp :: Lens' GetCoreDeviceResponse (Maybe UTCTime)

The time at which the core device's status last updated, expressed in ISO 8601 format.

getCoreDeviceResponse_httpStatus :: Lens' GetCoreDeviceResponse Int

The response's http status code.

GetComponentVersionArtifact

getComponentVersionArtifact_arn :: Lens' GetComponentVersionArtifact Text

The ARN of the component version. Specify the ARN of a public component version.

getComponentVersionArtifact_artifactName :: Lens' GetComponentVersionArtifact Text

The name of the artifact.

You can use the GetComponent operation to download the component recipe, which includes the URI of the artifact. The artifact name is the section of the URI after the scheme. For example, in the artifact URI greengrass:SomeArtifact.zip, the artifact name is SomeArtifact.zip.

getComponentVersionArtifactResponse_httpStatus :: Lens' GetComponentVersionArtifactResponse Int

The response's http status code.

getComponentVersionArtifactResponse_preSignedUrl :: Lens' GetComponentVersionArtifactResponse Text

The URL of the artifact.

CreateDeployment

createDeployment_components :: Lens' CreateDeployment (Maybe (HashMap Text ComponentDeploymentSpecification))

The components to deploy. This is a dictionary, where each key is the name of a component, and each key's value is the version and configuration to deploy for that component.

createDeployment_clientToken :: Lens' CreateDeployment (Maybe Text)

A unique, case-sensitive identifier that you can provide to ensure that the request is idempotent. Idempotency means that the request is successfully processed only once, even if you send the request multiple times. When a request succeeds, and you specify the same client token for subsequent successful requests, the IoT Greengrass V2 service returns the successful response that it caches from the previous request. IoT Greengrass V2 caches successful responses for idempotent requests for up to 8 hours.

createDeployment_deploymentPolicies :: Lens' CreateDeployment (Maybe DeploymentPolicies)

The deployment policies for the deployment. These policies define how the deployment updates components and handles failure.

createDeployment_iotJobConfiguration :: Lens' CreateDeployment (Maybe DeploymentIoTJobConfiguration)

The job configuration for the deployment configuration. The job configuration specifies the rollout, timeout, and stop configurations for the deployment configuration.

createDeployment_deploymentName :: Lens' CreateDeployment (Maybe Text)

The name of the deployment.

createDeployment_tags :: Lens' CreateDeployment (Maybe (HashMap Text Text))

A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

createDeployment_targetArn :: Lens' CreateDeployment Text

The ARN of the target IoT thing or thing group.

createDeploymentResponse_deploymentId :: Lens' CreateDeploymentResponse (Maybe Text)

The ID of the deployment.

createDeploymentResponse_iotJobId :: Lens' CreateDeploymentResponse (Maybe Text)

The ID of the IoT job that applies the deployment to target devices.

createDeploymentResponse_iotJobArn :: Lens' CreateDeploymentResponse (Maybe Text)

The ARN of the IoT job that applies the deployment to target devices.

createDeploymentResponse_httpStatus :: Lens' CreateDeploymentResponse Int

The response's http status code.

BatchAssociateClientDeviceWithCoreDevice

batchAssociateClientDeviceWithCoreDevice_entries :: Lens' BatchAssociateClientDeviceWithCoreDevice (Maybe (NonEmpty AssociateClientDeviceWithCoreDeviceEntry))

The list of client devices to associate.

batchAssociateClientDeviceWithCoreDevice_coreDeviceThingName :: Lens' BatchAssociateClientDeviceWithCoreDevice Text

The name of the core device. This is also the name of the IoT thing.

batchAssociateClientDeviceWithCoreDeviceResponse_errorEntries :: Lens' BatchAssociateClientDeviceWithCoreDeviceResponse (Maybe (NonEmpty AssociateClientDeviceWithCoreDeviceErrorEntry))

The list of any errors for the entries in the request. Each error entry contains the name of the IoT thing that failed to associate.

batchAssociateClientDeviceWithCoreDeviceResponse_httpStatus :: Lens' BatchAssociateClientDeviceWithCoreDeviceResponse Int

The response's http status code.

ListInstalledComponents

listInstalledComponents_nextToken :: Lens' ListInstalledComponents (Maybe Text)

The token to be used for the next set of paginated results.

listInstalledComponents_maxResults :: Lens' ListInstalledComponents (Maybe Natural)

The maximum number of results to be returned per paginated request.

listInstalledComponents_coreDeviceThingName :: Lens' ListInstalledComponents Text

The name of the core device. This is also the name of the IoT thing.

listInstalledComponentsResponse_installedComponents :: Lens' ListInstalledComponentsResponse (Maybe [InstalledComponent])

A list that summarizes each component on the core device.

listInstalledComponentsResponse_nextToken :: Lens' ListInstalledComponentsResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

listInstalledComponentsResponse_httpStatus :: Lens' ListInstalledComponentsResponse Int

The response's http status code.

CancelDeployment

cancelDeployment_deploymentId :: Lens' CancelDeployment Text

The ID of the deployment.

cancelDeploymentResponse_message :: Lens' CancelDeploymentResponse (Maybe Text)

A message that communicates if the cancel was successful.

cancelDeploymentResponse_httpStatus :: Lens' CancelDeploymentResponse Int

The response's http status code.

BatchDisassociateClientDeviceFromCoreDevice

batchDisassociateClientDeviceFromCoreDevice_entries :: Lens' BatchDisassociateClientDeviceFromCoreDevice (Maybe (NonEmpty DisassociateClientDeviceFromCoreDeviceEntry))

The list of client devices to disassociate.

batchDisassociateClientDeviceFromCoreDevice_coreDeviceThingName :: Lens' BatchDisassociateClientDeviceFromCoreDevice Text

The name of the core device. This is also the name of the IoT thing.

batchDisassociateClientDeviceFromCoreDeviceResponse_errorEntries :: Lens' BatchDisassociateClientDeviceFromCoreDeviceResponse (Maybe (NonEmpty DisassociateClientDeviceFromCoreDeviceErrorEntry))

The list of errors (if any) for the entries in the request. Each error entry contains the name of the IoT thing that failed to disassociate.

batchDisassociateClientDeviceFromCoreDeviceResponse_httpStatus :: Lens' BatchDisassociateClientDeviceFromCoreDeviceResponse Int

The response's http status code.

ListCoreDevices

listCoreDevices_status :: Lens' ListCoreDevices (Maybe CoreDeviceStatus)

The core device status by which to filter. If you specify this parameter, the list includes only core devices that have this status. Choose one of the following options:

HEALTHY – The IoT Greengrass Core software and all components run on the core device without issue.

UNHEALTHY – The IoT Greengrass Core software or a component is in a failed state on the core device.

listCoreDevices_thingGroupArn :: Lens' ListCoreDevices (Maybe Text)

The ARN of the IoT thing group by which to filter. If you specify this parameter, the list includes only core devices that are members of this thing group.

listCoreDevices_nextToken :: Lens' ListCoreDevices (Maybe Text)

The token to be used for the next set of paginated results.

listCoreDevices_maxResults :: Lens' ListCoreDevices (Maybe Natural)

The maximum number of results to be returned per paginated request.

listCoreDevicesResponse_nextToken :: Lens' ListCoreDevicesResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

listCoreDevicesResponse_coreDevices :: Lens' ListCoreDevicesResponse (Maybe [CoreDevice])

A list that summarizes each core device.

listCoreDevicesResponse_httpStatus :: Lens' ListCoreDevicesResponse Int

The response's http status code.

ResolveComponentCandidates

resolveComponentCandidates_platform :: Lens' ResolveComponentCandidates ComponentPlatform

The platform to use to resolve compatible components.

resolveComponentCandidates_componentCandidates :: Lens' ResolveComponentCandidates [ComponentCandidate]

The list of components to resolve.

resolveComponentCandidatesResponse_resolvedComponentVersions :: Lens' ResolveComponentCandidatesResponse (Maybe [ResolvedComponentVersion])

A list of components that meet the requirements that you specify in the request. This list includes each component's recipe that you can use to install the component.

resolveComponentCandidatesResponse_httpStatus :: Lens' ResolveComponentCandidatesResponse Int

The response's http status code.

ListEffectiveDeployments

listEffectiveDeployments_nextToken :: Lens' ListEffectiveDeployments (Maybe Text)

The token to be used for the next set of paginated results.

listEffectiveDeployments_maxResults :: Lens' ListEffectiveDeployments (Maybe Natural)

The maximum number of results to be returned per paginated request.

listEffectiveDeployments_coreDeviceThingName :: Lens' ListEffectiveDeployments Text

The name of the core device. This is also the name of the IoT thing.

listEffectiveDeploymentsResponse_nextToken :: Lens' ListEffectiveDeploymentsResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

listEffectiveDeploymentsResponse_effectiveDeployments :: Lens' ListEffectiveDeploymentsResponse (Maybe [EffectiveDeployment])

A list that summarizes each deployment on the core device.

listEffectiveDeploymentsResponse_httpStatus :: Lens' ListEffectiveDeploymentsResponse Int

The response's http status code.

ListDeployments

listDeployments_targetArn :: Lens' ListDeployments (Maybe Text)

The ARN of the target IoT thing or thing group.

listDeployments_nextToken :: Lens' ListDeployments (Maybe Text)

The token to be used for the next set of paginated results.

listDeployments_historyFilter :: Lens' ListDeployments (Maybe DeploymentHistoryFilter)

The filter for the list of deployments. Choose one of the following options:

ALL – The list includes all deployments.

LATEST_ONLY – The list includes only the latest revision of each deployment.

Default: LATEST_ONLY

listDeployments_maxResults :: Lens' ListDeployments (Maybe Natural)

The maximum number of results to be returned per paginated request.

listDeploymentsResponse_nextToken :: Lens' ListDeploymentsResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

listDeploymentsResponse_deployments :: Lens' ListDeploymentsResponse (Maybe [Deployment])

A list that summarizes each deployment.

listDeploymentsResponse_httpStatus :: Lens' ListDeploymentsResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The ARN of the resource to tag.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The ARN of the resource to untag.

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

A list of keys for tags to remove from the resource.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

ListComponents

listComponents_nextToken :: Lens' ListComponents (Maybe Text)

The token to be used for the next set of paginated results.

listComponents_scope :: Lens' ListComponents (Maybe ComponentVisibilityScope)

The scope of the components to list.

Default: PRIVATE

listComponents_maxResults :: Lens' ListComponents (Maybe Natural)

The maximum number of results to be returned per paginated request.

listComponentsResponse_components :: Lens' ListComponentsResponse (Maybe [Component])

A list that summarizes each component.

listComponentsResponse_nextToken :: Lens' ListComponentsResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

listComponentsResponse_httpStatus :: Lens' ListComponentsResponse Int

The response's http status code.

DeleteComponent

deleteComponent_arn :: Lens' DeleteComponent Text

The ARN of the component version.

CreateComponentVersion

createComponentVersion_clientToken :: Lens' CreateComponentVersion (Maybe Text)

A unique, case-sensitive identifier that you can provide to ensure that the request is idempotent. Idempotency means that the request is successfully processed only once, even if you send the request multiple times. When a request succeeds, and you specify the same client token for subsequent successful requests, the IoT Greengrass V2 service returns the successful response that it caches from the previous request. IoT Greengrass V2 caches successful responses for idempotent requests for up to 8 hours.

createComponentVersion_inlineRecipe :: Lens' CreateComponentVersion (Maybe ByteString)

The recipe to use to create the component. The recipe defines the component's metadata, parameters, dependencies, lifecycle, artifacts, and platform compatibility.

You must specify either inlineRecipe or lambdaFunction.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

createComponentVersion_lambdaFunction :: Lens' CreateComponentVersion (Maybe LambdaFunctionRecipeSource)

The parameters to create a component from a Lambda function.

You must specify either inlineRecipe or lambdaFunction.

createComponentVersion_tags :: Lens' CreateComponentVersion (Maybe (HashMap Text Text))

A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

createComponentVersionResponse_arn :: Lens' CreateComponentVersionResponse (Maybe Text)

The ARN of the component version.

createComponentVersionResponse_httpStatus :: Lens' CreateComponentVersionResponse Int

The response's http status code.

createComponentVersionResponse_componentName :: Lens' CreateComponentVersionResponse Text

The name of the component.

createComponentVersionResponse_componentVersion :: Lens' CreateComponentVersionResponse Text

The version of the component.

createComponentVersionResponse_creationTimestamp :: Lens' CreateComponentVersionResponse UTCTime

The time at which the component was created, expressed in ISO 8601 format.

createComponentVersionResponse_status :: Lens' CreateComponentVersionResponse CloudComponentStatus

The status of the component version in IoT Greengrass V2. This status is different from the status of the component on a core device.

GetComponent

getComponent_recipeOutputFormat :: Lens' GetComponent (Maybe RecipeOutputFormat)

The format of the recipe.

getComponent_arn :: Lens' GetComponent Text

The ARN of the component version.

getComponentResponse_tags :: Lens' GetComponentResponse (Maybe (HashMap Text Text))

A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

getComponentResponse_httpStatus :: Lens' GetComponentResponse Int

The response's http status code.

getComponentResponse_recipeOutputFormat :: Lens' GetComponentResponse RecipeOutputFormat

The format of the recipe.

getComponentResponse_recipe :: Lens' GetComponentResponse ByteString

The recipe of the component version.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

ListClientDevicesAssociatedWithCoreDevice

listClientDevicesAssociatedWithCoreDevice_nextToken :: Lens' ListClientDevicesAssociatedWithCoreDevice (Maybe Text)

The token to be used for the next set of paginated results.

listClientDevicesAssociatedWithCoreDevice_maxResults :: Lens' ListClientDevicesAssociatedWithCoreDevice (Maybe Natural)

The maximum number of results to be returned per paginated request.

listClientDevicesAssociatedWithCoreDevice_coreDeviceThingName :: Lens' ListClientDevicesAssociatedWithCoreDevice Text

The name of the core device. This is also the name of the IoT thing.

listClientDevicesAssociatedWithCoreDeviceResponse_associatedClientDevices :: Lens' ListClientDevicesAssociatedWithCoreDeviceResponse (Maybe (NonEmpty AssociatedClientDevice))

A list that describes the client devices that are associated with the core device.

listClientDevicesAssociatedWithCoreDeviceResponse_nextToken :: Lens' ListClientDevicesAssociatedWithCoreDeviceResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

listClientDevicesAssociatedWithCoreDeviceResponse_httpStatus :: Lens' ListClientDevicesAssociatedWithCoreDeviceResponse Int

The response's http status code.

Types

AssociateClientDeviceWithCoreDeviceEntry

associateClientDeviceWithCoreDeviceEntry_thingName :: Lens' AssociateClientDeviceWithCoreDeviceEntry Text

The name of the IoT thing that represents the client device to associate.

AssociateClientDeviceWithCoreDeviceErrorEntry

associateClientDeviceWithCoreDeviceErrorEntry_code :: Lens' AssociateClientDeviceWithCoreDeviceErrorEntry (Maybe Text)

The error code for the request.

associateClientDeviceWithCoreDeviceErrorEntry_message :: Lens' AssociateClientDeviceWithCoreDeviceErrorEntry (Maybe Text)

A message that provides additional information about the error.

associateClientDeviceWithCoreDeviceErrorEntry_thingName :: Lens' AssociateClientDeviceWithCoreDeviceErrorEntry (Maybe Text)

The name of the IoT thing whose associate request failed.

AssociatedClientDevice

associatedClientDevice_associationTimestamp :: Lens' AssociatedClientDevice (Maybe UTCTime)

The time that the client device was associated, expressed in ISO 8601 format.

associatedClientDevice_thingName :: Lens' AssociatedClientDevice (Maybe Text)

The name of the IoT thing that represents the associated client device.

CloudComponentStatus

cloudComponentStatus_componentState :: Lens' CloudComponentStatus (Maybe CloudComponentState)

The state of the component.

cloudComponentStatus_message :: Lens' CloudComponentStatus (Maybe Text)

A message that communicates details, such as errors, about the status of the component.

cloudComponentStatus_errors :: Lens' CloudComponentStatus (Maybe (HashMap Text Text))

A dictionary of errors that communicate why the component is in an error state. For example, if IoT Greengrass can't access an artifact for the component, then errors contains the artifact's URI as a key, and the error message as the value for that key.

Component

component_arn :: Lens' Component (Maybe Text)

The ARN of the component version.

component_componentName :: Lens' Component (Maybe Text)

The name of the component.

component_latestVersion :: Lens' Component (Maybe ComponentLatestVersion)

The latest version of the component and its details.

ComponentCandidate

componentCandidate_componentVersion :: Lens' ComponentCandidate (Maybe Text)

The version of the component.

componentCandidate_versionRequirements :: Lens' ComponentCandidate (Maybe (HashMap Text Text))

The version requirements for the component's dependencies. Greengrass core devices get the version requirements from component recipes.

IoT Greengrass V2 uses semantic version constraints. For more information, see Semantic Versioning.

componentCandidate_componentName :: Lens' ComponentCandidate (Maybe Text)

The name of the component.

ComponentConfigurationUpdate

componentConfigurationUpdate_reset :: Lens' ComponentConfigurationUpdate (Maybe [Text])

The list of configuration nodes to reset to default values on target devices. Use JSON pointers to specify each node to reset. JSON pointers start with a forward slash (/) and use forward slashes to separate the key for each level in the object. For more information, see the JSON pointer specification and Reset configuration updates in the IoT Greengrass V2 Developer Guide.

componentConfigurationUpdate_merge :: Lens' ComponentConfigurationUpdate (Maybe Text)

A serialized JSON string that contains the configuration object to merge to target devices. The core device merges this configuration with the component's existing configuration. If this is the first time a component deploys on a device, the core device merges this configuration with the component's default configuration. This means that the core device keeps it's existing configuration for keys and values that you don't specify in this object. For more information, see Merge configuration updates in the IoT Greengrass V2 Developer Guide.

ComponentDependencyRequirement

componentDependencyRequirement_dependencyType :: Lens' ComponentDependencyRequirement (Maybe ComponentDependencyType)

The type of this dependency. Choose from the following options:

SOFT – The component doesn't restart if the dependency changes state.

HARD – The component restarts if the dependency changes state.

Default: HARD

componentDependencyRequirement_versionRequirement :: Lens' ComponentDependencyRequirement (Maybe Text)

The component version requirement for the component dependency.

IoT Greengrass V2 uses semantic version constraints. For more information, see Semantic Versioning.

ComponentDeploymentSpecification

componentDeploymentSpecification_componentVersion :: Lens' ComponentDeploymentSpecification (Maybe Text)

The version of the component.

componentDeploymentSpecification_runWith :: Lens' ComponentDeploymentSpecification (Maybe ComponentRunWith)

The system user and group that the IoT Greengrass Core software uses to run component processes on the core device. If you omit this parameter, the IoT Greengrass Core software uses the system user and group that you configure for the core device. For more information, see Configure the user and group that run components in the IoT Greengrass V2 Developer Guide.

componentDeploymentSpecification_configurationUpdate :: Lens' ComponentDeploymentSpecification (Maybe ComponentConfigurationUpdate)

The configuration updates to deploy for the component. You can define reset updates and merge updates. A reset updates the keys that you specify to the default configuration for the component. A merge updates the core device's component configuration with the keys and values that you specify. The IoT Greengrass Core software applies reset updates before it applies merge updates. For more information, see Update component configurations in the IoT Greengrass V2 Developer Guide.

ComponentLatestVersion

componentLatestVersion_platforms :: Lens' ComponentLatestVersion (Maybe [ComponentPlatform])

The platforms that the component version supports.

componentLatestVersion_arn :: Lens' ComponentLatestVersion (Maybe Text)

The ARN of the component version.

componentLatestVersion_componentVersion :: Lens' ComponentLatestVersion (Maybe Text)

The version of the component.

componentLatestVersion_creationTimestamp :: Lens' ComponentLatestVersion (Maybe UTCTime)

The time at which the component was created, expressed in ISO 8601 format.

componentLatestVersion_publisher :: Lens' ComponentLatestVersion (Maybe Text)

The publisher of the component version.

componentLatestVersion_description :: Lens' ComponentLatestVersion (Maybe Text)

The description of the component version.

ComponentPlatform

componentPlatform_name :: Lens' ComponentPlatform (Maybe Text)

The friendly name of the platform. This name helps you identify the platform.

If you omit this parameter, IoT Greengrass creates a friendly name from the os and architecture of the platform.

componentPlatform_attributes :: Lens' ComponentPlatform (Maybe (HashMap Text Text))

A dictionary of attributes for the platform. The IoT Greengrass Core software defines the os and platform by default. You can specify additional platform attributes for a core device when you deploy the Greengrass nucleus component. For more information, see the Greengrass nucleus component in the IoT Greengrass V2 Developer Guide.

ComponentRunWith

componentRunWith_posixUser :: Lens' ComponentRunWith (Maybe Text)

The POSIX system user and (optional) group to use to run this component. Specify the user and group separated by a colon (:) in the following format: user:group. The group is optional. If you don't specify a group, the IoT Greengrass Core software uses the primary user for the group.

If you omit this parameter, the IoT Greengrass Core software uses the default system user and group that you configure on the Greengrass nucleus component. For more information, see Configure the user and group that run components.

componentRunWith_systemResourceLimits :: Lens' ComponentRunWith (Maybe SystemResourceLimits)

The system resource limits to apply to this component's process on the core device.

If you omit this parameter, the IoT Greengrass Core software uses the default system resource limits that you configure on the Greengrass nucleus component. For more information, see Configure system resource limits for components.

ComponentVersionListItem

componentVersionListItem_arn :: Lens' ComponentVersionListItem (Maybe Text)

The ARN of the component version.

componentVersionListItem_componentVersion :: Lens' ComponentVersionListItem (Maybe Text)

The version of the component.

componentVersionListItem_componentName :: Lens' ComponentVersionListItem (Maybe Text)

The name of the component.

CoreDevice

coreDevice_status :: Lens' CoreDevice (Maybe CoreDeviceStatus)

The status of the core device. Core devices can have the following statuses:

HEALTHY – The IoT Greengrass Core software and all components run on the core device without issue.

UNHEALTHY – The IoT Greengrass Core software or a component is in a failed state on the core device.

coreDevice_coreDeviceThingName :: Lens' CoreDevice (Maybe Text)

The name of the core device. This is also the name of the IoT thing.

coreDevice_lastStatusUpdateTimestamp :: Lens' CoreDevice (Maybe UTCTime)

The time at which the core device's status last updated, expressed in ISO 8601 format.

Deployment

deployment_targetArn :: Lens' Deployment (Maybe Text)

The ARN of the target IoT thing or thing group.

deployment_deploymentId :: Lens' Deployment (Maybe Text)

The ID of the deployment.

deployment_creationTimestamp :: Lens' Deployment (Maybe UTCTime)

The time at which the deployment was created, expressed in ISO 8601 format.

deployment_deploymentStatus :: Lens' Deployment (Maybe DeploymentStatus)

The status of the deployment.

deployment_isLatestForTarget :: Lens' Deployment (Maybe Bool)

Whether or not the deployment is the latest revision for its target.

deployment_revisionId :: Lens' Deployment (Maybe Text)

The revision number of the deployment.

deployment_deploymentName :: Lens' Deployment (Maybe Text)

The name of the deployment.

DeploymentComponentUpdatePolicy

deploymentComponentUpdatePolicy_action :: Lens' DeploymentComponentUpdatePolicy (Maybe DeploymentComponentUpdatePolicyAction)

Whether or not to notify components and wait for components to become safe to update. Choose from the following options:

NOTIFY_COMPONENTS – The deployment notifies each component before it stops and updates that component. Components can use the SubscribeToComponentUpdates IPC operation to receive these notifications. Then, components can respond with the DeferComponentUpdate IPC operation. For more information, see Create deployments in the IoT Greengrass V2 Developer Guide.

SKIP_NOTIFY_COMPONENTS – The deployment doesn't notify components or wait for them to be safe to update.

Default: NOTIFY_COMPONENTS

deploymentComponentUpdatePolicy_timeoutInSeconds :: Lens' DeploymentComponentUpdatePolicy (Maybe Int)

The amount of time in seconds that each component on a device has to report that it's safe to update. If the component waits for longer than this timeout, then the deployment proceeds on the device.

Default: 60

DeploymentConfigurationValidationPolicy

deploymentConfigurationValidationPolicy_timeoutInSeconds :: Lens' DeploymentConfigurationValidationPolicy (Maybe Int)

The amount of time in seconds that a component can validate its configuration updates. If the validation time exceeds this timeout, then the deployment proceeds for the device.

Default: 30

DeploymentIoTJobConfiguration

deploymentIoTJobConfiguration_jobExecutionsRolloutConfig :: Lens' DeploymentIoTJobConfiguration (Maybe IoTJobExecutionsRolloutConfig)

The rollout configuration for the job. This configuration defines the rate at which the job rolls out to the fleet of target devices.

deploymentIoTJobConfiguration_abortConfig :: Lens' DeploymentIoTJobConfiguration (Maybe IoTJobAbortConfig)

The stop configuration for the job. This configuration defines when and how to stop a job rollout.

deploymentIoTJobConfiguration_timeoutConfig :: Lens' DeploymentIoTJobConfiguration (Maybe IoTJobTimeoutConfig)

The timeout configuration for the job. This configuration defines the amount of time each device has to complete the job.

DeploymentPolicies

deploymentPolicies_failureHandlingPolicy :: Lens' DeploymentPolicies (Maybe DeploymentFailureHandlingPolicy)

The failure handling policy for the configuration deployment. This policy defines what to do if the deployment fails.

Default: ROLLBACK

deploymentPolicies_configurationValidationPolicy :: Lens' DeploymentPolicies (Maybe DeploymentConfigurationValidationPolicy)

The configuration validation policy for the configuration deployment. This policy defines how long each component has to validate its configure updates.

deploymentPolicies_componentUpdatePolicy :: Lens' DeploymentPolicies (Maybe DeploymentComponentUpdatePolicy)

The component update policy for the configuration deployment. This policy defines when it's safe to deploy the configuration to devices.

DisassociateClientDeviceFromCoreDeviceEntry

disassociateClientDeviceFromCoreDeviceEntry_thingName :: Lens' DisassociateClientDeviceFromCoreDeviceEntry Text

The name of the IoT thing that represents the client device to disassociate.

DisassociateClientDeviceFromCoreDeviceErrorEntry

disassociateClientDeviceFromCoreDeviceErrorEntry_code :: Lens' DisassociateClientDeviceFromCoreDeviceErrorEntry (Maybe Text)

The error code for the request.

disassociateClientDeviceFromCoreDeviceErrorEntry_message :: Lens' DisassociateClientDeviceFromCoreDeviceErrorEntry (Maybe Text)

A message that provides additional information about the error.

disassociateClientDeviceFromCoreDeviceErrorEntry_thingName :: Lens' DisassociateClientDeviceFromCoreDeviceErrorEntry (Maybe Text)

The name of the IoT thing whose disassociate request failed.

EffectiveDeployment

effectiveDeployment_iotJobId :: Lens' EffectiveDeployment (Maybe Text)

The ID of the IoT job that applies the deployment to target devices.

effectiveDeployment_iotJobArn :: Lens' EffectiveDeployment (Maybe Text)

The ARN of the IoT job that applies the deployment to target devices.

effectiveDeployment_reason :: Lens' EffectiveDeployment (Maybe Text)

The reason code for the update, if the job was updated.

effectiveDeployment_description :: Lens' EffectiveDeployment (Maybe Text)

The description of the deployment job.

effectiveDeployment_deploymentId :: Lens' EffectiveDeployment Text

The ID of the deployment.

effectiveDeployment_deploymentName :: Lens' EffectiveDeployment Text

The name of the deployment.

effectiveDeployment_targetArn :: Lens' EffectiveDeployment Text

The ARN of the target IoT thing or thing group.

effectiveDeployment_coreDeviceExecutionStatus :: Lens' EffectiveDeployment EffectiveDeploymentExecutionStatus

The status of the deployment job on the Greengrass core device.

effectiveDeployment_creationTimestamp :: Lens' EffectiveDeployment UTCTime

The time at which the deployment was created, expressed in ISO 8601 format.

effectiveDeployment_modifiedTimestamp :: Lens' EffectiveDeployment UTCTime

The time at which the deployment job was last modified, expressed in ISO 8601 format.

InstalledComponent

installedComponent_isRoot :: Lens' InstalledComponent (Maybe Bool)

Whether or not the component is a root component.

installedComponent_componentVersion :: Lens' InstalledComponent (Maybe Text)

The version of the component.

installedComponent_componentName :: Lens' InstalledComponent (Maybe Text)

The name of the component.

installedComponent_lifecycleStateDetails :: Lens' InstalledComponent (Maybe Text)

The details about the lifecycle state of the component.

installedComponent_lifecycleState :: Lens' InstalledComponent (Maybe InstalledComponentLifecycleState)

The lifecycle state of the component.

IoTJobAbortConfig

ioTJobAbortConfig_criteriaList :: Lens' IoTJobAbortConfig (NonEmpty IoTJobAbortCriteria)

The list of criteria that define when and how to cancel the configuration deployment.

IoTJobAbortCriteria

ioTJobAbortCriteria_failureType :: Lens' IoTJobAbortCriteria IoTJobExecutionFailureType

The type of job deployment failure that can cancel a job.

ioTJobAbortCriteria_action :: Lens' IoTJobAbortCriteria IoTJobAbortAction

The action to perform when the criteria are met.

ioTJobAbortCriteria_thresholdPercentage :: Lens' IoTJobAbortCriteria Double

The minimum percentage of failureType failures that occur before the job can cancel.

This parameter supports up to two digits after the decimal (for example, you can specify 10.9 or 10.99, but not 10.999).

ioTJobAbortCriteria_minNumberOfExecutedThings :: Lens' IoTJobAbortCriteria Natural

The minimum number of things that receive the configuration before the job can cancel.

IoTJobExecutionsRolloutConfig

ioTJobExecutionsRolloutConfig_exponentialRate :: Lens' IoTJobExecutionsRolloutConfig (Maybe IoTJobExponentialRolloutRate)

The exponential rate to increase the job rollout rate.

ioTJobExecutionsRolloutConfig_maximumPerMinute :: Lens' IoTJobExecutionsRolloutConfig (Maybe Natural)

The maximum number of devices that receive a pending job notification, per minute.

IoTJobExponentialRolloutRate

ioTJobExponentialRolloutRate_baseRatePerMinute :: Lens' IoTJobExponentialRolloutRate Natural

The minimum number of devices that receive a pending job notification, per minute, when the job starts. This parameter defines the initial rollout rate of the job.

ioTJobExponentialRolloutRate_incrementFactor :: Lens' IoTJobExponentialRolloutRate Double

The exponential factor to increase the rollout rate for the job.

This parameter supports up to one digit after the decimal (for example, you can specify 1.5, but not 1.55).

ioTJobExponentialRolloutRate_rateIncreaseCriteria :: Lens' IoTJobExponentialRolloutRate IoTJobRateIncreaseCriteria

The criteria to increase the rollout rate for the job.

IoTJobRateIncreaseCriteria

ioTJobRateIncreaseCriteria_numberOfNotifiedThings :: Lens' IoTJobRateIncreaseCriteria (Maybe Natural)

The number of devices to receive the job notification before the rollout rate increases.

ioTJobRateIncreaseCriteria_numberOfSucceededThings :: Lens' IoTJobRateIncreaseCriteria (Maybe Natural)

The number of devices to successfully run the configuration job before the rollout rate increases.

IoTJobTimeoutConfig

ioTJobTimeoutConfig_inProgressTimeoutInMinutes :: Lens' IoTJobTimeoutConfig (Maybe Integer)

The amount of time, in minutes, that devices have to complete the job. The timer starts when the job status is set to IN_PROGRESS. If the job status doesn't change to a terminal state before the time expires, then the job status is set to TIMED_OUT.

The timeout interval must be between 1 minute and 7 days (10080 minutes).

LambdaContainerParams

lambdaContainerParams_mountROSysfs :: Lens' LambdaContainerParams (Maybe Bool)

Whether or not the container can read information from the device's /sys folder.

Default: false

lambdaContainerParams_memorySizeInKB :: Lens' LambdaContainerParams (Maybe Int)

The memory size of the container, expressed in kilobytes.

Default: 16384 (16 MB)

lambdaContainerParams_devices :: Lens' LambdaContainerParams (Maybe [LambdaDeviceMount])

The list of system devices that the container can access.

lambdaContainerParams_volumes :: Lens' LambdaContainerParams (Maybe [LambdaVolumeMount])

The list of volumes that the container can access.

LambdaDeviceMount

lambdaDeviceMount_addGroupOwner :: Lens' LambdaDeviceMount (Maybe Bool)

Whether or not to add the component's system user as an owner of the device.

Default: false

lambdaDeviceMount_permission :: Lens' LambdaDeviceMount (Maybe LambdaFilesystemPermission)

The permission to access the device: read/only (ro) or read/write (rw).

Default: ro

lambdaDeviceMount_path :: Lens' LambdaDeviceMount Text

The mount path for the device in the file system.

LambdaEventSource

lambdaEventSource_topic :: Lens' LambdaEventSource Text

The topic to which to subscribe to receive event messages.

lambdaEventSource_type :: Lens' LambdaEventSource LambdaEventSourceType

The type of event source. Choose from the following options:

PUB_SUB – Subscribe to local publish/subscribe messages. This event source type doesn't support MQTT wildcards (+ and #) in the event source topic.

IOT_CORE – Subscribe to Amazon Web Services IoT Core MQTT messages. This event source type supports MQTT wildcards (+ and #) in the event source topic.

LambdaExecutionParameters

lambdaExecutionParameters_execArgs :: Lens' LambdaExecutionParameters (Maybe [Text])

The list of arguments to pass to the Lambda function when it runs.

lambdaExecutionParameters_maxQueueSize :: Lens' LambdaExecutionParameters (Maybe Int)

The maximum size of the message queue for the Lambda function component. The IoT Greengrass core stores messages in a FIFO (first-in-first-out) queue until it can run the Lambda function to consume each message.

lambdaExecutionParameters_pinned :: Lens' LambdaExecutionParameters (Maybe Bool)

Whether or not the Lambda function is pinned, or long-lived.

A pinned Lambda function starts when IoT Greengrass starts and keeps running in its own container.

A non-pinned Lambda function starts only when it receives a work item and exists after it idles for maxIdleTimeInSeconds. If the function has multiple work items, the IoT Greengrass Core software creates multiple instances of the function.

Default: true

lambdaExecutionParameters_inputPayloadEncodingType :: Lens' LambdaExecutionParameters (Maybe LambdaInputPayloadEncodingType)

The encoding type that the Lambda function supports.

Default: json

lambdaExecutionParameters_maxIdleTimeInSeconds :: Lens' LambdaExecutionParameters (Maybe Int)

The maximum amount of time in seconds that a non-pinned Lambda function can idle before the IoT Greengrass Core software stops its process.

lambdaExecutionParameters_timeoutInSeconds :: Lens' LambdaExecutionParameters (Maybe Int)

The maximum amount of time in seconds that the Lambda function can process a work item.

lambdaExecutionParameters_eventSources :: Lens' LambdaExecutionParameters (Maybe [LambdaEventSource])

The list of event sources to which to subscribe to receive work messages. The Lambda function runs when it receives a message from an event source. You can subscribe this function to local publish/subscribe messages and Amazon Web Services IoT Core MQTT messages.

lambdaExecutionParameters_maxInstancesCount :: Lens' LambdaExecutionParameters (Maybe Int)

The maximum number of instances that a non-pinned Lambda function can run at the same time.

lambdaExecutionParameters_environmentVariables :: Lens' LambdaExecutionParameters (Maybe (HashMap Text Text))

The map of environment variables that are available to the Lambda function when it runs.

lambdaExecutionParameters_statusTimeoutInSeconds :: Lens' LambdaExecutionParameters (Maybe Int)

The interval in seconds at which a pinned (also known as long-lived) Lambda function component sends status updates to the Lambda manager component.

lambdaExecutionParameters_linuxProcessParams :: Lens' LambdaExecutionParameters (Maybe LambdaLinuxProcessParams)

The parameters for the Linux process that contains the Lambda function.

LambdaFunctionRecipeSource

lambdaFunctionRecipeSource_componentLambdaParameters :: Lens' LambdaFunctionRecipeSource (Maybe LambdaExecutionParameters)

The system and runtime parameters for the Lambda function as it runs on the Greengrass core device.

lambdaFunctionRecipeSource_componentVersion :: Lens' LambdaFunctionRecipeSource (Maybe Text)

The version of the component.

Defaults to the version of the Lambda function as a semantic version. For example, if your function version is 3, the component version becomes 3.0.0.

lambdaFunctionRecipeSource_componentName :: Lens' LambdaFunctionRecipeSource (Maybe Text)

The name of the component.

Defaults to the name of the Lambda function.

lambdaFunctionRecipeSource_componentPlatforms :: Lens' LambdaFunctionRecipeSource (Maybe [ComponentPlatform])

The platforms that the component version supports.

lambdaFunctionRecipeSource_componentDependencies :: Lens' LambdaFunctionRecipeSource (Maybe (HashMap Text ComponentDependencyRequirement))

The component versions on which this Lambda function component depends.

lambdaFunctionRecipeSource_lambdaArn :: Lens' LambdaFunctionRecipeSource Text

The ARN of the Lambda function. The ARN must include the version of the function to import. You can't use version aliases like $LATEST.

LambdaLinuxProcessParams

lambdaLinuxProcessParams_containerParams :: Lens' LambdaLinuxProcessParams (Maybe LambdaContainerParams)

The parameters for the container in which the Lambda function runs.

lambdaLinuxProcessParams_isolationMode :: Lens' LambdaLinuxProcessParams (Maybe LambdaIsolationMode)

The isolation mode for the process that contains the Lambda function. The process can run in an isolated runtime environment inside the IoT Greengrass container, or as a regular process outside any container.

Default: GreengrassContainer

LambdaVolumeMount

lambdaVolumeMount_addGroupOwner :: Lens' LambdaVolumeMount (Maybe Bool)

Whether or not to add the IoT Greengrass user group as an owner of the volume.

Default: false

lambdaVolumeMount_permission :: Lens' LambdaVolumeMount (Maybe LambdaFilesystemPermission)

The permission to access the volume: read/only (ro) or read/write (rw).

Default: ro

lambdaVolumeMount_sourcePath :: Lens' LambdaVolumeMount Text

The path to the physical volume in the file system.

lambdaVolumeMount_destinationPath :: Lens' LambdaVolumeMount Text

The path to the logical volume in the file system.

ResolvedComponentVersion

resolvedComponentVersion_arn :: Lens' ResolvedComponentVersion (Maybe Text)

The ARN of the component version.

resolvedComponentVersion_componentVersion :: Lens' ResolvedComponentVersion (Maybe Text)

The version of the component.

resolvedComponentVersion_recipe :: Lens' ResolvedComponentVersion (Maybe ByteString)

The recipe of the component version.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

resolvedComponentVersion_componentName :: Lens' ResolvedComponentVersion (Maybe Text)

The name of the component.

SystemResourceLimits

systemResourceLimits_memory :: Lens' SystemResourceLimits (Maybe Natural)

The maximum amount of RAM, expressed in kilobytes, that a component's processes can use on the core device.

systemResourceLimits_cpus :: Lens' SystemResourceLimits (Maybe Double)

The maximum amount of CPU time that a component's processes can use on the core device. A core device's total CPU time is equivalent to the device's number of CPU cores. For example, on a core device with 4 CPU cores, you can set this value to 2 to limit the component's processes to 50 percent usage of each CPU core. On a device with 1 CPU core, you can set this value to 0.25 to limit the component's processes to 25 percent usage of the CPU. If you set this value to a number greater than the number of CPU cores, the IoT Greengrass Core software doesn't limit the component's CPU usage.