Network.AWS.GreengrassV2.ListClientDevicesAssociatedWithCoreDevice

Description

Retrieves a paginated list of client devices that are associated with a core device.

This operation returns paginated results.

Creating a Request

data ListClientDevicesAssociatedWithCoreDevice

See: newListClientDevicesAssociatedWithCoreDevice smart constructor.

Constructors

ListClientDevicesAssociatedWithCoreDevice'

Fields

nextToken :: Maybe Text

The token to be used for the next set of paginated results.

maxResults :: Maybe Natural

The maximum number of results to be returned per paginated request.

coreDeviceThingName :: Text

The name of the core device. This is also the name of the IoT thing.

newListClientDevicesAssociatedWithCoreDevice

Arguments

:: Text
$sel:coreDeviceThingName:ListClientDevicesAssociatedWithCoreDevice'

-> ListClientDevicesAssociatedWithCoreDevice

Create a value of ListClientDevicesAssociatedWithCoreDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListClientDevicesAssociatedWithCoreDevice', listClientDevicesAssociatedWithCoreDevice_nextToken - The token to be used for the next set of paginated results.

$sel:maxResults:ListClientDevicesAssociatedWithCoreDevice', listClientDevicesAssociatedWithCoreDevice_maxResults - The maximum number of results to be returned per paginated request.

$sel:coreDeviceThingName:ListClientDevicesAssociatedWithCoreDevice', listClientDevicesAssociatedWithCoreDevice_coreDeviceThingName - The name of the core device. This is also the name of the IoT thing.

Request Lenses

listClientDevicesAssociatedWithCoreDevice_nextToken :: Lens' ListClientDevicesAssociatedWithCoreDevice (Maybe Text)

The token to be used for the next set of paginated results.

listClientDevicesAssociatedWithCoreDevice_maxResults :: Lens' ListClientDevicesAssociatedWithCoreDevice (Maybe Natural)

The maximum number of results to be returned per paginated request.

listClientDevicesAssociatedWithCoreDevice_coreDeviceThingName :: Lens' ListClientDevicesAssociatedWithCoreDevice Text

The name of the core device. This is also the name of the IoT thing.

Destructuring the Response

data ListClientDevicesAssociatedWithCoreDeviceResponse

See: newListClientDevicesAssociatedWithCoreDeviceResponse smart constructor.

Constructors

ListClientDevicesAssociatedWithCoreDeviceResponse'

Fields

associatedClientDevices :: Maybe (NonEmpty AssociatedClientDevice)

A list that describes the client devices that are associated with the core device.

nextToken :: Maybe Text

The token for the next set of results, or null if there are no additional results.

httpStatus :: Int

The response's http status code.

newListClientDevicesAssociatedWithCoreDeviceResponse

Arguments

:: Int
$sel:httpStatus:ListClientDevicesAssociatedWithCoreDeviceResponse'

-> ListClientDevicesAssociatedWithCoreDeviceResponse

Create a value of ListClientDevicesAssociatedWithCoreDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associatedClientDevices:ListClientDevicesAssociatedWithCoreDeviceResponse', listClientDevicesAssociatedWithCoreDeviceResponse_associatedClientDevices - A list that describes the client devices that are associated with the core device.

$sel:nextToken:ListClientDevicesAssociatedWithCoreDevice', listClientDevicesAssociatedWithCoreDeviceResponse_nextToken - The token for the next set of results, or null if there are no additional results.

$sel:httpStatus:ListClientDevicesAssociatedWithCoreDeviceResponse', listClientDevicesAssociatedWithCoreDeviceResponse_httpStatus - The response's http status code.

Response Lenses

listClientDevicesAssociatedWithCoreDeviceResponse_associatedClientDevices :: Lens' ListClientDevicesAssociatedWithCoreDeviceResponse (Maybe (NonEmpty AssociatedClientDevice))

A list that describes the client devices that are associated with the core device.

listClientDevicesAssociatedWithCoreDeviceResponse_nextToken :: Lens' ListClientDevicesAssociatedWithCoreDeviceResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

listClientDevicesAssociatedWithCoreDeviceResponse_httpStatus :: Lens' ListClientDevicesAssociatedWithCoreDeviceResponse Int

The response's http status code.