Network.AWS.GreengrassV2.Types.CloudComponentStatus

Description

Documentation

data CloudComponentStatus

Contains the status of a component in the IoT Greengrass service.

See: newCloudComponentStatus smart constructor.

Constructors

CloudComponentStatus'

Fields

componentState :: Maybe CloudComponentState

The state of the component.

message :: Maybe Text

A message that communicates details, such as errors, about the status of the component.

errors :: Maybe (HashMap Text Text)

A dictionary of errors that communicate why the component is in an error state. For example, if IoT Greengrass can't access an artifact for the component, then errors contains the artifact's URI as a key, and the error message as the value for that key.

newCloudComponentStatus :: CloudComponentStatus

Create a value of CloudComponentStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:componentState:CloudComponentStatus', cloudComponentStatus_componentState - The state of the component.

$sel:message:CloudComponentStatus', cloudComponentStatus_message - A message that communicates details, such as errors, about the status of the component.

$sel:errors:CloudComponentStatus', cloudComponentStatus_errors - A dictionary of errors that communicate why the component is in an error state. For example, if IoT Greengrass can't access an artifact for the component, then errors contains the artifact's URI as a key, and the error message as the value for that key.

cloudComponentStatus_componentState :: Lens' CloudComponentStatus (Maybe CloudComponentState)

The state of the component.

cloudComponentStatus_message :: Lens' CloudComponentStatus (Maybe Text)

A message that communicates details, such as errors, about the status of the component.

cloudComponentStatus_errors :: Lens' CloudComponentStatus (Maybe (HashMap Text Text))

A dictionary of errors that communicate why the component is in an error state. For example, if IoT Greengrass can't access an artifact for the component, then errors contains the artifact's URI as a key, and the error message as the value for that key.