Network.AWS.GreengrassV2.Types.Component

Description

Documentation

data Component

Contains information about a component.

See: newComponent smart constructor.

Constructors

Component'

Fields

arn :: Maybe Text

The ARN of the component version.

componentName :: Maybe Text

The name of the component.

latestVersion :: Maybe ComponentLatestVersion

The latest version of the component and its details.

newComponent :: Component

Create a value of Component with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Component', component_arn - The ARN of the component version.

$sel:componentName:Component', component_componentName - The name of the component.

$sel:latestVersion:Component', component_latestVersion - The latest version of the component and its details.

component_arn :: Lens' Component (Maybe Text)

The ARN of the component version.

component_componentName :: Lens' Component (Maybe Text)

The name of the component.

component_latestVersion :: Lens' Component (Maybe ComponentLatestVersion)

The latest version of the component and its details.