Network.AWS.GreengrassV2.Types.ComponentConfigurationUpdate

Description

Documentation

data ComponentConfigurationUpdate

Contains information about a deployment's update to a component's configuration on Greengrass core devices. For more information, see Update component configurations in the IoT Greengrass V2 Developer Guide.

See: newComponentConfigurationUpdate smart constructor.

Constructors

ComponentConfigurationUpdate'

Fields

reset :: Maybe [Text]

The list of configuration nodes to reset to default values on target devices. Use JSON pointers to specify each node to reset. JSON pointers start with a forward slash (/) and use forward slashes to separate the key for each level in the object. For more information, see the JSON pointer specification and Reset configuration updates in the IoT Greengrass V2 Developer Guide.

merge :: Maybe Text

A serialized JSON string that contains the configuration object to merge to target devices. The core device merges this configuration with the component's existing configuration. If this is the first time a component deploys on a device, the core device merges this configuration with the component's default configuration. This means that the core device keeps it's existing configuration for keys and values that you don't specify in this object. For more information, see Merge configuration updates in the IoT Greengrass V2 Developer Guide.

newComponentConfigurationUpdate :: ComponentConfigurationUpdate

Create a value of ComponentConfigurationUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reset:ComponentConfigurationUpdate', componentConfigurationUpdate_reset - The list of configuration nodes to reset to default values on target devices. Use JSON pointers to specify each node to reset. JSON pointers start with a forward slash (/) and use forward slashes to separate the key for each level in the object. For more information, see the JSON pointer specification and Reset configuration updates in the IoT Greengrass V2 Developer Guide.

$sel:merge:ComponentConfigurationUpdate', componentConfigurationUpdate_merge - A serialized JSON string that contains the configuration object to merge to target devices. The core device merges this configuration with the component's existing configuration. If this is the first time a component deploys on a device, the core device merges this configuration with the component's default configuration. This means that the core device keeps it's existing configuration for keys and values that you don't specify in this object. For more information, see Merge configuration updates in the IoT Greengrass V2 Developer Guide.

componentConfigurationUpdate_reset :: Lens' ComponentConfigurationUpdate (Maybe [Text])

The list of configuration nodes to reset to default values on target devices. Use JSON pointers to specify each node to reset. JSON pointers start with a forward slash (/) and use forward slashes to separate the key for each level in the object. For more information, see the JSON pointer specification and Reset configuration updates in the IoT Greengrass V2 Developer Guide.

componentConfigurationUpdate_merge :: Lens' ComponentConfigurationUpdate (Maybe Text)

A serialized JSON string that contains the configuration object to merge to target devices. The core device merges this configuration with the component's existing configuration. If this is the first time a component deploys on a device, the core device merges this configuration with the component's default configuration. This means that the core device keeps it's existing configuration for keys and values that you don't specify in this object. For more information, see Merge configuration updates in the IoT Greengrass V2 Developer Guide.