Network.AWS.GreengrassV2.Types.ComponentDeploymentSpecification

Description

Documentation

data ComponentDeploymentSpecification

Contains information about a component to deploy.

See: newComponentDeploymentSpecification smart constructor.

Constructors

ComponentDeploymentSpecification'

Fields

componentVersion :: Maybe Text

The version of the component.

runWith :: Maybe ComponentRunWith

The system user and group that the IoT Greengrass Core software uses to run component processes on the core device. If you omit this parameter, the IoT Greengrass Core software uses the system user and group that you configure for the core device. For more information, see Configure the user and group that run components in the IoT Greengrass V2 Developer Guide.

configurationUpdate :: Maybe ComponentConfigurationUpdate

The configuration updates to deploy for the component. You can define reset updates and merge updates. A reset updates the keys that you specify to the default configuration for the component. A merge updates the core device's component configuration with the keys and values that you specify. The IoT Greengrass Core software applies reset updates before it applies merge updates. For more information, see Update component configurations in the IoT Greengrass V2 Developer Guide.

newComponentDeploymentSpecification :: ComponentDeploymentSpecification

Create a value of ComponentDeploymentSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:componentVersion:ComponentDeploymentSpecification', componentDeploymentSpecification_componentVersion - The version of the component.

$sel:runWith:ComponentDeploymentSpecification', componentDeploymentSpecification_runWith - The system user and group that the IoT Greengrass Core software uses to run component processes on the core device. If you omit this parameter, the IoT Greengrass Core software uses the system user and group that you configure for the core device. For more information, see Configure the user and group that run components in the IoT Greengrass V2 Developer Guide.

$sel:configurationUpdate:ComponentDeploymentSpecification', componentDeploymentSpecification_configurationUpdate - The configuration updates to deploy for the component. You can define reset updates and merge updates. A reset updates the keys that you specify to the default configuration for the component. A merge updates the core device's component configuration with the keys and values that you specify. The IoT Greengrass Core software applies reset updates before it applies merge updates. For more information, see Update component configurations in the IoT Greengrass V2 Developer Guide.

componentDeploymentSpecification_componentVersion :: Lens' ComponentDeploymentSpecification (Maybe Text)

The version of the component.

componentDeploymentSpecification_runWith :: Lens' ComponentDeploymentSpecification (Maybe ComponentRunWith)

The system user and group that the IoT Greengrass Core software uses to run component processes on the core device. If you omit this parameter, the IoT Greengrass Core software uses the system user and group that you configure for the core device. For more information, see Configure the user and group that run components in the IoT Greengrass V2 Developer Guide.

componentDeploymentSpecification_configurationUpdate :: Lens' ComponentDeploymentSpecification (Maybe ComponentConfigurationUpdate)

The configuration updates to deploy for the component. You can define reset updates and merge updates. A reset updates the keys that you specify to the default configuration for the component. A merge updates the core device's component configuration with the keys and values that you specify. The IoT Greengrass Core software applies reset updates before it applies merge updates. For more information, see Update component configurations in the IoT Greengrass V2 Developer Guide.