Network.AWS.GreengrassV2.Types.ComponentRunWith

Description

Documentation

data ComponentRunWith

Contains information system user and group that the IoT Greengrass Core software uses to run component processes on the core device. For more information, see Configure the user and group that run components in the IoT Greengrass V2 Developer Guide.

See: newComponentRunWith smart constructor.

Constructors

ComponentRunWith'

Fields

posixUser :: Maybe Text

The POSIX system user and (optional) group to use to run this component. Specify the user and group separated by a colon (:) in the following format: user:group. The group is optional. If you don't specify a group, the IoT Greengrass Core software uses the primary user for the group.

If you omit this parameter, the IoT Greengrass Core software uses the default system user and group that you configure on the Greengrass nucleus component. For more information, see Configure the user and group that run components.

systemResourceLimits :: Maybe SystemResourceLimits

The system resource limits to apply to this component's process on the core device.

If you omit this parameter, the IoT Greengrass Core software uses the default system resource limits that you configure on the Greengrass nucleus component. For more information, see Configure system resource limits for components.

newComponentRunWith :: ComponentRunWith

Create a value of ComponentRunWith with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:posixUser:ComponentRunWith', componentRunWith_posixUser - The POSIX system user and (optional) group to use to run this component. Specify the user and group separated by a colon (:) in the following format: user:group. The group is optional. If you don't specify a group, the IoT Greengrass Core software uses the primary user for the group.

If you omit this parameter, the IoT Greengrass Core software uses the default system user and group that you configure on the Greengrass nucleus component. For more information, see Configure the user and group that run components.

$sel:systemResourceLimits:ComponentRunWith', componentRunWith_systemResourceLimits - The system resource limits to apply to this component's process on the core device.

If you omit this parameter, the IoT Greengrass Core software uses the default system resource limits that you configure on the Greengrass nucleus component. For more information, see Configure system resource limits for components.

componentRunWith_posixUser :: Lens' ComponentRunWith (Maybe Text)

The POSIX system user and (optional) group to use to run this component. Specify the user and group separated by a colon (:) in the following format: user:group. The group is optional. If you don't specify a group, the IoT Greengrass Core software uses the primary user for the group.

If you omit this parameter, the IoT Greengrass Core software uses the default system user and group that you configure on the Greengrass nucleus component. For more information, see Configure the user and group that run components.

componentRunWith_systemResourceLimits :: Lens' ComponentRunWith (Maybe SystemResourceLimits)

The system resource limits to apply to this component's process on the core device.

If you omit this parameter, the IoT Greengrass Core software uses the default system resource limits that you configure on the Greengrass nucleus component. For more information, see Configure system resource limits for components.