Network.AWS.GreengrassV2.Types.DeploymentConfigurationValidationPolicy

Description

Documentation

data DeploymentConfigurationValidationPolicy

Contains information about how long a component on a core device can validate its configuration updates before it times out. Components can use the SubscribeToValidateConfigurationUpdates IPC operation to receive notifications when a deployment specifies a configuration update. Then, components can respond with the SendConfigurationValidityReport IPC operation. For more information, see Create deployments in the IoT Greengrass V2 Developer Guide.

See: newDeploymentConfigurationValidationPolicy smart constructor.

Constructors

DeploymentConfigurationValidationPolicy'

Fields

timeoutInSeconds :: Maybe Int

The amount of time in seconds that a component can validate its configuration updates. If the validation time exceeds this timeout, then the deployment proceeds for the device.

Default: 30

newDeploymentConfigurationValidationPolicy :: DeploymentConfigurationValidationPolicy

Create a value of DeploymentConfigurationValidationPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeoutInSeconds:DeploymentConfigurationValidationPolicy', deploymentConfigurationValidationPolicy_timeoutInSeconds - The amount of time in seconds that a component can validate its configuration updates. If the validation time exceeds this timeout, then the deployment proceeds for the device.

Default: 30

deploymentConfigurationValidationPolicy_timeoutInSeconds :: Lens' DeploymentConfigurationValidationPolicy (Maybe Int)

The amount of time in seconds that a component can validate its configuration updates. If the validation time exceeds this timeout, then the deployment proceeds for the device.

Default: 30