Network.AWS.GreengrassV2.Types.EffectiveDeployment

Description

Documentation

data EffectiveDeployment

Contains information about a deployment job that IoT Greengrass sends to a Greengrass core device.

See: newEffectiveDeployment smart constructor.

Constructors

EffectiveDeployment'

Fields

iotJobId :: Maybe Text

The ID of the IoT job that applies the deployment to target devices.

iotJobArn :: Maybe Text

The ARN of the IoT job that applies the deployment to target devices.

reason :: Maybe Text

The reason code for the update, if the job was updated.

description :: Maybe Text

The description of the deployment job.

deploymentId :: Text

The ID of the deployment.

deploymentName :: Text

The name of the deployment.

targetArn :: Text

The ARN of the target IoT thing or thing group.

coreDeviceExecutionStatus :: EffectiveDeploymentExecutionStatus

The status of the deployment job on the Greengrass core device.

creationTimestamp :: POSIX

The time at which the deployment was created, expressed in ISO 8601 format.

modifiedTimestamp :: POSIX

The time at which the deployment job was last modified, expressed in ISO 8601 format.

newEffectiveDeployment

Arguments

:: Text
$sel:deploymentId:EffectiveDeployment'

-> Text
$sel:deploymentName:EffectiveDeployment'

-> Text
$sel:targetArn:EffectiveDeployment'

-> EffectiveDeploymentExecutionStatus
$sel:coreDeviceExecutionStatus:EffectiveDeployment'

-> UTCTime
$sel:creationTimestamp:EffectiveDeployment'

-> UTCTime
$sel:modifiedTimestamp:EffectiveDeployment'

-> EffectiveDeployment

Create a value of EffectiveDeployment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iotJobId:EffectiveDeployment', effectiveDeployment_iotJobId - The ID of the IoT job that applies the deployment to target devices.

$sel:iotJobArn:EffectiveDeployment', effectiveDeployment_iotJobArn - The ARN of the IoT job that applies the deployment to target devices.

$sel:reason:EffectiveDeployment', effectiveDeployment_reason - The reason code for the update, if the job was updated.

$sel:description:EffectiveDeployment', effectiveDeployment_description - The description of the deployment job.

$sel:deploymentId:EffectiveDeployment', effectiveDeployment_deploymentId - The ID of the deployment.

$sel:deploymentName:EffectiveDeployment', effectiveDeployment_deploymentName - The name of the deployment.

$sel:targetArn:EffectiveDeployment', effectiveDeployment_targetArn - The ARN of the target IoT thing or thing group.

$sel:coreDeviceExecutionStatus:EffectiveDeployment', effectiveDeployment_coreDeviceExecutionStatus - The status of the deployment job on the Greengrass core device.

$sel:creationTimestamp:EffectiveDeployment', effectiveDeployment_creationTimestamp - The time at which the deployment was created, expressed in ISO 8601 format.

$sel:modifiedTimestamp:EffectiveDeployment', effectiveDeployment_modifiedTimestamp - The time at which the deployment job was last modified, expressed in ISO 8601 format.

effectiveDeployment_iotJobId :: Lens' EffectiveDeployment (Maybe Text)

The ID of the IoT job that applies the deployment to target devices.

effectiveDeployment_iotJobArn :: Lens' EffectiveDeployment (Maybe Text)

The ARN of the IoT job that applies the deployment to target devices.

effectiveDeployment_reason :: Lens' EffectiveDeployment (Maybe Text)

The reason code for the update, if the job was updated.

effectiveDeployment_description :: Lens' EffectiveDeployment (Maybe Text)

The description of the deployment job.

effectiveDeployment_deploymentId :: Lens' EffectiveDeployment Text

The ID of the deployment.

effectiveDeployment_deploymentName :: Lens' EffectiveDeployment Text

The name of the deployment.

effectiveDeployment_targetArn :: Lens' EffectiveDeployment Text

The ARN of the target IoT thing or thing group.

effectiveDeployment_coreDeviceExecutionStatus :: Lens' EffectiveDeployment EffectiveDeploymentExecutionStatus

The status of the deployment job on the Greengrass core device.

effectiveDeployment_creationTimestamp :: Lens' EffectiveDeployment UTCTime

The time at which the deployment was created, expressed in ISO 8601 format.

effectiveDeployment_modifiedTimestamp :: Lens' EffectiveDeployment UTCTime

The time at which the deployment job was last modified, expressed in ISO 8601 format.