Network.AWS.GreengrassV2.Types.LambdaExecutionParameters

Description

Documentation

data LambdaExecutionParameters

Contains parameters for a Lambda function that runs on IoT Greengrass.

See: newLambdaExecutionParameters smart constructor.

Constructors

LambdaExecutionParameters'

Fields

execArgs :: Maybe [Text]

The list of arguments to pass to the Lambda function when it runs.

maxQueueSize :: Maybe Int

The maximum size of the message queue for the Lambda function component. The IoT Greengrass core stores messages in a FIFO (first-in-first-out) queue until it can run the Lambda function to consume each message.

pinned :: Maybe Bool

Whether or not the Lambda function is pinned, or long-lived.

A pinned Lambda function starts when IoT Greengrass starts and keeps running in its own container.

A non-pinned Lambda function starts only when it receives a work item and exists after it idles for maxIdleTimeInSeconds. If the function has multiple work items, the IoT Greengrass Core software creates multiple instances of the function.

Default: true

inputPayloadEncodingType :: Maybe LambdaInputPayloadEncodingType

The encoding type that the Lambda function supports.

Default: json

maxIdleTimeInSeconds :: Maybe Int

The maximum amount of time in seconds that a non-pinned Lambda function can idle before the IoT Greengrass Core software stops its process.

timeoutInSeconds :: Maybe Int

The maximum amount of time in seconds that the Lambda function can process a work item.

eventSources :: Maybe [LambdaEventSource]

The list of event sources to which to subscribe to receive work messages. The Lambda function runs when it receives a message from an event source. You can subscribe this function to local publish/subscribe messages and Amazon Web Services IoT Core MQTT messages.

maxInstancesCount :: Maybe Int

The maximum number of instances that a non-pinned Lambda function can run at the same time.

environmentVariables :: Maybe (HashMap Text Text)

The map of environment variables that are available to the Lambda function when it runs.

statusTimeoutInSeconds :: Maybe Int

The interval in seconds at which a pinned (also known as long-lived) Lambda function component sends status updates to the Lambda manager component.

linuxProcessParams :: Maybe LambdaLinuxProcessParams

The parameters for the Linux process that contains the Lambda function.

newLambdaExecutionParameters :: LambdaExecutionParameters

Create a value of LambdaExecutionParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:execArgs:LambdaExecutionParameters', lambdaExecutionParameters_execArgs - The list of arguments to pass to the Lambda function when it runs.

$sel:maxQueueSize:LambdaExecutionParameters', lambdaExecutionParameters_maxQueueSize - The maximum size of the message queue for the Lambda function component. The IoT Greengrass core stores messages in a FIFO (first-in-first-out) queue until it can run the Lambda function to consume each message.

$sel:pinned:LambdaExecutionParameters', lambdaExecutionParameters_pinned - Whether or not the Lambda function is pinned, or long-lived.

A pinned Lambda function starts when IoT Greengrass starts and keeps running in its own container.

A non-pinned Lambda function starts only when it receives a work item and exists after it idles for maxIdleTimeInSeconds. If the function has multiple work items, the IoT Greengrass Core software creates multiple instances of the function.

Default: true

$sel:inputPayloadEncodingType:LambdaExecutionParameters', lambdaExecutionParameters_inputPayloadEncodingType - The encoding type that the Lambda function supports.

Default: json

$sel:maxIdleTimeInSeconds:LambdaExecutionParameters', lambdaExecutionParameters_maxIdleTimeInSeconds - The maximum amount of time in seconds that a non-pinned Lambda function can idle before the IoT Greengrass Core software stops its process.

$sel:timeoutInSeconds:LambdaExecutionParameters', lambdaExecutionParameters_timeoutInSeconds - The maximum amount of time in seconds that the Lambda function can process a work item.

$sel:eventSources:LambdaExecutionParameters', lambdaExecutionParameters_eventSources - The list of event sources to which to subscribe to receive work messages. The Lambda function runs when it receives a message from an event source. You can subscribe this function to local publish/subscribe messages and Amazon Web Services IoT Core MQTT messages.

$sel:maxInstancesCount:LambdaExecutionParameters', lambdaExecutionParameters_maxInstancesCount - The maximum number of instances that a non-pinned Lambda function can run at the same time.

$sel:environmentVariables:LambdaExecutionParameters', lambdaExecutionParameters_environmentVariables - The map of environment variables that are available to the Lambda function when it runs.

$sel:statusTimeoutInSeconds:LambdaExecutionParameters', lambdaExecutionParameters_statusTimeoutInSeconds - The interval in seconds at which a pinned (also known as long-lived) Lambda function component sends status updates to the Lambda manager component.

$sel:linuxProcessParams:LambdaExecutionParameters', lambdaExecutionParameters_linuxProcessParams - The parameters for the Linux process that contains the Lambda function.

lambdaExecutionParameters_execArgs :: Lens' LambdaExecutionParameters (Maybe [Text])

The list of arguments to pass to the Lambda function when it runs.

lambdaExecutionParameters_maxQueueSize :: Lens' LambdaExecutionParameters (Maybe Int)

The maximum size of the message queue for the Lambda function component. The IoT Greengrass core stores messages in a FIFO (first-in-first-out) queue until it can run the Lambda function to consume each message.

lambdaExecutionParameters_pinned :: Lens' LambdaExecutionParameters (Maybe Bool)

Whether or not the Lambda function is pinned, or long-lived.

A pinned Lambda function starts when IoT Greengrass starts and keeps running in its own container.

A non-pinned Lambda function starts only when it receives a work item and exists after it idles for maxIdleTimeInSeconds. If the function has multiple work items, the IoT Greengrass Core software creates multiple instances of the function.

Default: true

lambdaExecutionParameters_inputPayloadEncodingType :: Lens' LambdaExecutionParameters (Maybe LambdaInputPayloadEncodingType)

The encoding type that the Lambda function supports.

Default: json

lambdaExecutionParameters_maxIdleTimeInSeconds :: Lens' LambdaExecutionParameters (Maybe Int)

The maximum amount of time in seconds that a non-pinned Lambda function can idle before the IoT Greengrass Core software stops its process.

lambdaExecutionParameters_timeoutInSeconds :: Lens' LambdaExecutionParameters (Maybe Int)

The maximum amount of time in seconds that the Lambda function can process a work item.

lambdaExecutionParameters_eventSources :: Lens' LambdaExecutionParameters (Maybe [LambdaEventSource])

The list of event sources to which to subscribe to receive work messages. The Lambda function runs when it receives a message from an event source. You can subscribe this function to local publish/subscribe messages and Amazon Web Services IoT Core MQTT messages.

lambdaExecutionParameters_maxInstancesCount :: Lens' LambdaExecutionParameters (Maybe Int)

The maximum number of instances that a non-pinned Lambda function can run at the same time.

lambdaExecutionParameters_environmentVariables :: Lens' LambdaExecutionParameters (Maybe (HashMap Text Text))

The map of environment variables that are available to the Lambda function when it runs.

lambdaExecutionParameters_statusTimeoutInSeconds :: Lens' LambdaExecutionParameters (Maybe Int)

The interval in seconds at which a pinned (also known as long-lived) Lambda function component sends status updates to the Lambda manager component.

lambdaExecutionParameters_linuxProcessParams :: Lens' LambdaExecutionParameters (Maybe LambdaLinuxProcessParams)

The parameters for the Linux process that contains the Lambda function.