Network.AWS.GreengrassV2.Types.SystemResourceLimits

Description

Documentation

data SystemResourceLimits

Contains information about system resource limits that the IoT Greengrass Core software applies to a component's processes. For more information, see Configure system resource limits for components.

See: newSystemResourceLimits smart constructor.

Constructors

SystemResourceLimits'

Fields

memory :: Maybe Natural

The maximum amount of RAM, expressed in kilobytes, that a component's processes can use on the core device.

cpus :: Maybe Double

The maximum amount of CPU time that a component's processes can use on the core device. A core device's total CPU time is equivalent to the device's number of CPU cores. For example, on a core device with 4 CPU cores, you can set this value to 2 to limit the component's processes to 50 percent usage of each CPU core. On a device with 1 CPU core, you can set this value to 0.25 to limit the component's processes to 25 percent usage of the CPU. If you set this value to a number greater than the number of CPU cores, the IoT Greengrass Core software doesn't limit the component's CPU usage.

newSystemResourceLimits :: SystemResourceLimits

Create a value of SystemResourceLimits with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memory:SystemResourceLimits', systemResourceLimits_memory - The maximum amount of RAM, expressed in kilobytes, that a component's processes can use on the core device.

$sel:cpus:SystemResourceLimits', systemResourceLimits_cpus - The maximum amount of CPU time that a component's processes can use on the core device. A core device's total CPU time is equivalent to the device's number of CPU cores. For example, on a core device with 4 CPU cores, you can set this value to 2 to limit the component's processes to 50 percent usage of each CPU core. On a device with 1 CPU core, you can set this value to 0.25 to limit the component's processes to 25 percent usage of the CPU. If you set this value to a number greater than the number of CPU cores, the IoT Greengrass Core software doesn't limit the component's CPU usage.

systemResourceLimits_memory :: Lens' SystemResourceLimits (Maybe Natural)

The maximum amount of RAM, expressed in kilobytes, that a component's processes can use on the core device.

systemResourceLimits_cpus :: Lens' SystemResourceLimits (Maybe Double)

The maximum amount of CPU time that a component's processes can use on the core device. A core device's total CPU time is equivalent to the device's number of CPU cores. For example, on a core device with 4 CPU cores, you can set this value to 2 to limit the component's processes to 50 percent usage of each CPU core. On a device with 1 CPU core, you can set this value to 0.25 to limit the component's processes to 25 percent usage of the CPU. If you set this value to a number greater than the number of CPU cores, the IoT Greengrass Core software doesn't limit the component's CPU usage.