Network.AWS.GroundStation.CreateDataflowEndpointGroup

Description

Creates a DataflowEndpoint group containing the specified list of DataflowEndpoint objects.

The name field in each endpoint is used in your mission profile DataflowEndpointConfig to specify which endpoints to use during a contact.

When a contact uses multiple DataflowEndpointConfig objects, each Config must match a DataflowEndpoint in the same group.

Creating a Request

data CreateDataflowEndpointGroup

See: newCreateDataflowEndpointGroup smart constructor.

Constructors

CreateDataflowEndpointGroup'

Fields

tags :: Maybe (HashMap Text Text)

Tags of a dataflow endpoint group.

endpointDetails :: [EndpointDetails]

Endpoint details of each endpoint in the dataflow endpoint group.

newCreateDataflowEndpointGroup :: CreateDataflowEndpointGroup

Create a value of CreateDataflowEndpointGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateDataflowEndpointGroup', createDataflowEndpointGroup_tags - Tags of a dataflow endpoint group.

$sel:endpointDetails:CreateDataflowEndpointGroup', createDataflowEndpointGroup_endpointDetails - Endpoint details of each endpoint in the dataflow endpoint group.

Request Lenses

createDataflowEndpointGroup_tags :: Lens' CreateDataflowEndpointGroup (Maybe (HashMap Text Text))

Tags of a dataflow endpoint group.

createDataflowEndpointGroup_endpointDetails :: Lens' CreateDataflowEndpointGroup [EndpointDetails]

Endpoint details of each endpoint in the dataflow endpoint group.

Destructuring the Response

data DataflowEndpointGroupIdResponse

See: newDataflowEndpointGroupIdResponse smart constructor.

Constructors

DataflowEndpointGroupIdResponse'

Fields

dataflowEndpointGroupId :: Maybe Text

UUID of a dataflow endpoint group.

newDataflowEndpointGroupIdResponse :: DataflowEndpointGroupIdResponse

Create a value of DataflowEndpointGroupIdResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataflowEndpointGroupId:DataflowEndpointGroupIdResponse', dataflowEndpointGroupIdResponse_dataflowEndpointGroupId - UUID of a dataflow endpoint group.

Response Lenses

dataflowEndpointGroupIdResponse_dataflowEndpointGroupId :: Lens' DataflowEndpointGroupIdResponse (Maybe Text)

UUID of a dataflow endpoint group.