Network.AWS.GroundStation.DeleteConfig

Description

Deletes a Config.

Creating a Request

data DeleteConfig

See: newDeleteConfig smart constructor.

Constructors

DeleteConfig'

Fields

configId :: Text

UUID of a Config.

configType :: ConfigCapabilityType

Type of a Config.

newDeleteConfig

Arguments

:: Text
$sel:configId:DeleteConfig'

-> ConfigCapabilityType
$sel:configType:DeleteConfig'

-> DeleteConfig

Create a value of DeleteConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configId:DeleteConfig', deleteConfig_configId - UUID of a Config.

$sel:configType:DeleteConfig', deleteConfig_configType - Type of a Config.

Request Lenses

deleteConfig_configId :: Lens' DeleteConfig Text

UUID of a Config.

deleteConfig_configType :: Lens' DeleteConfig ConfigCapabilityType

Type of a Config.

Destructuring the Response

data ConfigIdResponse

See: newConfigIdResponse smart constructor.

Constructors

ConfigIdResponse'

Fields

configArn :: Maybe Text

ARN of a Config.

configId :: Maybe Text

UUID of a Config.

configType :: Maybe ConfigCapabilityType

Type of a Config.

newConfigIdResponse :: ConfigIdResponse

Create a value of ConfigIdResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configArn:ConfigIdResponse', configIdResponse_configArn - ARN of a Config.

$sel:configId:ConfigIdResponse', configIdResponse_configId - UUID of a Config.

$sel:configType:ConfigIdResponse', configIdResponse_configType - Type of a Config.

Response Lenses

configIdResponse_configArn :: Lens' ConfigIdResponse (Maybe Text)

ARN of a Config.

configIdResponse_configId :: Lens' ConfigIdResponse (Maybe Text)

UUID of a Config.

configIdResponse_configType :: Lens' ConfigIdResponse (Maybe ConfigCapabilityType)

Type of a Config.