Network.AWS.GroundStation.DescribeContact

Description

Describes an existing contact.

Creating a Request

data DescribeContact

See: newDescribeContact smart constructor.

Constructors

DescribeContact'

Fields

contactId :: Text

UUID of a contact.

newDescribeContact

Arguments

:: Text
$sel:contactId:DescribeContact'

-> DescribeContact

Create a value of DescribeContact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contactId:DescribeContact', describeContact_contactId - UUID of a contact.

Request Lenses

describeContact_contactId :: Lens' DescribeContact Text

UUID of a contact.

Destructuring the Response

data DescribeContactResponse

See: newDescribeContactResponse smart constructor.

Constructors

DescribeContactResponse'

Fields

contactStatus :: Maybe ContactStatus

Status of a contact.

missionProfileArn :: Maybe Text

ARN of a mission profile.

startTime :: Maybe POSIX

Start time of a contact.

dataflowList :: Maybe [DataflowDetail]

List describing source and destination details for each dataflow edge.

satelliteArn :: Maybe Text

ARN of a satellite.

maximumElevation :: Maybe Elevation

Maximum elevation angle of a contact.

groundStation :: Maybe Text

Ground station for a contact.

endTime :: Maybe POSIX

End time of a contact.

contactId :: Maybe Text

UUID of a contact.

region :: Maybe Text

Region of a contact.

postPassEndTime :: Maybe POSIX

Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

prePassStartTime :: Maybe POSIX

Amount of time prior to contact start you’d like to receive a CloudWatch event indicating an upcoming pass.

errorMessage :: Maybe Text

Error message for a contact.

tags :: Maybe (HashMap Text Text)

Tags assigned to a contact.

httpStatus :: Int

The response's http status code.

newDescribeContactResponse

Arguments

:: Int
$sel:httpStatus:DescribeContactResponse'

-> DescribeContactResponse

Create a value of DescribeContactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contactStatus:DescribeContactResponse', describeContactResponse_contactStatus - Status of a contact.

$sel:missionProfileArn:DescribeContactResponse', describeContactResponse_missionProfileArn - ARN of a mission profile.

$sel:startTime:DescribeContactResponse', describeContactResponse_startTime - Start time of a contact.

$sel:dataflowList:DescribeContactResponse', describeContactResponse_dataflowList - List describing source and destination details for each dataflow edge.

$sel:satelliteArn:DescribeContactResponse', describeContactResponse_satelliteArn - ARN of a satellite.

$sel:maximumElevation:DescribeContactResponse', describeContactResponse_maximumElevation - Maximum elevation angle of a contact.

$sel:groundStation:DescribeContactResponse', describeContactResponse_groundStation - Ground station for a contact.

$sel:endTime:DescribeContactResponse', describeContactResponse_endTime - End time of a contact.

$sel:contactId:DescribeContact', describeContactResponse_contactId - UUID of a contact.

$sel:region:DescribeContactResponse', describeContactResponse_region - Region of a contact.

$sel:postPassEndTime:DescribeContactResponse', describeContactResponse_postPassEndTime - Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

$sel:prePassStartTime:DescribeContactResponse', describeContactResponse_prePassStartTime - Amount of time prior to contact start you’d like to receive a CloudWatch event indicating an upcoming pass.

$sel:errorMessage:DescribeContactResponse', describeContactResponse_errorMessage - Error message for a contact.

$sel:tags:DescribeContactResponse', describeContactResponse_tags - Tags assigned to a contact.

$sel:httpStatus:DescribeContactResponse', describeContactResponse_httpStatus - The response's http status code.

Response Lenses

describeContactResponse_contactStatus :: Lens' DescribeContactResponse (Maybe ContactStatus)

Status of a contact.

describeContactResponse_missionProfileArn :: Lens' DescribeContactResponse (Maybe Text)

ARN of a mission profile.

describeContactResponse_startTime :: Lens' DescribeContactResponse (Maybe UTCTime)

Start time of a contact.

describeContactResponse_dataflowList :: Lens' DescribeContactResponse (Maybe [DataflowDetail])

List describing source and destination details for each dataflow edge.

describeContactResponse_satelliteArn :: Lens' DescribeContactResponse (Maybe Text)

ARN of a satellite.

describeContactResponse_maximumElevation :: Lens' DescribeContactResponse (Maybe Elevation)

Maximum elevation angle of a contact.

describeContactResponse_groundStation :: Lens' DescribeContactResponse (Maybe Text)

Ground station for a contact.

describeContactResponse_endTime :: Lens' DescribeContactResponse (Maybe UTCTime)

End time of a contact.

describeContactResponse_contactId :: Lens' DescribeContactResponse (Maybe Text)

UUID of a contact.

describeContactResponse_region :: Lens' DescribeContactResponse (Maybe Text)

Region of a contact.

describeContactResponse_postPassEndTime :: Lens' DescribeContactResponse (Maybe UTCTime)

Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

describeContactResponse_prePassStartTime :: Lens' DescribeContactResponse (Maybe UTCTime)

Amount of time prior to contact start you’d like to receive a CloudWatch event indicating an upcoming pass.

describeContactResponse_errorMessage :: Lens' DescribeContactResponse (Maybe Text)

Error message for a contact.

describeContactResponse_tags :: Lens' DescribeContactResponse (Maybe (HashMap Text Text))

Tags assigned to a contact.

describeContactResponse_httpStatus :: Lens' DescribeContactResponse Int

The response's http status code.