Network.AWS.GroundStation.GetDataflowEndpointGroup

Description

Returns the dataflow endpoint group.

Creating a Request

data GetDataflowEndpointGroup

See: newGetDataflowEndpointGroup smart constructor.

Constructors

GetDataflowEndpointGroup'

Fields

dataflowEndpointGroupId :: Text

UUID of a dataflow endpoint group.

newGetDataflowEndpointGroup

Arguments

:: Text
$sel:dataflowEndpointGroupId:GetDataflowEndpointGroup'

-> GetDataflowEndpointGroup

Create a value of GetDataflowEndpointGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataflowEndpointGroupId:GetDataflowEndpointGroup', getDataflowEndpointGroup_dataflowEndpointGroupId - UUID of a dataflow endpoint group.

Request Lenses

getDataflowEndpointGroup_dataflowEndpointGroupId :: Lens' GetDataflowEndpointGroup Text

UUID of a dataflow endpoint group.

Destructuring the Response

data GetDataflowEndpointGroupResponse

See: newGetDataflowEndpointGroupResponse smart constructor.

Constructors

GetDataflowEndpointGroupResponse'

Fields

endpointsDetails :: Maybe [EndpointDetails]

Details of a dataflow endpoint.

dataflowEndpointGroupArn :: Maybe Text

ARN of a dataflow endpoint group.

dataflowEndpointGroupId :: Maybe Text

UUID of a dataflow endpoint group.

tags :: Maybe (HashMap Text Text)

Tags assigned to a dataflow endpoint group.

httpStatus :: Int

The response's http status code.

newGetDataflowEndpointGroupResponse

Arguments

:: Int
$sel:httpStatus:GetDataflowEndpointGroupResponse'

-> GetDataflowEndpointGroupResponse

Create a value of GetDataflowEndpointGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpointsDetails:GetDataflowEndpointGroupResponse', getDataflowEndpointGroupResponse_endpointsDetails - Details of a dataflow endpoint.

$sel:dataflowEndpointGroupArn:GetDataflowEndpointGroupResponse', getDataflowEndpointGroupResponse_dataflowEndpointGroupArn - ARN of a dataflow endpoint group.

$sel:dataflowEndpointGroupId:GetDataflowEndpointGroup', getDataflowEndpointGroupResponse_dataflowEndpointGroupId - UUID of a dataflow endpoint group.

$sel:tags:GetDataflowEndpointGroupResponse', getDataflowEndpointGroupResponse_tags - Tags assigned to a dataflow endpoint group.

$sel:httpStatus:GetDataflowEndpointGroupResponse', getDataflowEndpointGroupResponse_httpStatus - The response's http status code.

Response Lenses

getDataflowEndpointGroupResponse_endpointsDetails :: Lens' GetDataflowEndpointGroupResponse (Maybe [EndpointDetails])

Details of a dataflow endpoint.

getDataflowEndpointGroupResponse_dataflowEndpointGroupArn :: Lens' GetDataflowEndpointGroupResponse (Maybe Text)

ARN of a dataflow endpoint group.

getDataflowEndpointGroupResponse_dataflowEndpointGroupId :: Lens' GetDataflowEndpointGroupResponse (Maybe Text)

UUID of a dataflow endpoint group.

getDataflowEndpointGroupResponse_tags :: Lens' GetDataflowEndpointGroupResponse (Maybe (HashMap Text Text))

Tags assigned to a dataflow endpoint group.

getDataflowEndpointGroupResponse_httpStatus :: Lens' GetDataflowEndpointGroupResponse Int

The response's http status code.