Network.AWS.GroundStation.GetMissionProfile

Description

Returns a mission profile.

Creating a Request

data GetMissionProfile

See: newGetMissionProfile smart constructor.

Constructors

GetMissionProfile'

Fields

missionProfileId :: Text

UUID of a mission profile.

newGetMissionProfile

Arguments

:: Text
$sel:missionProfileId:GetMissionProfile'

-> GetMissionProfile

Create a value of GetMissionProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:missionProfileId:GetMissionProfile', getMissionProfile_missionProfileId - UUID of a mission profile.

Request Lenses

getMissionProfile_missionProfileId :: Lens' GetMissionProfile Text

UUID of a mission profile.

Destructuring the Response

data GetMissionProfileResponse

See: newGetMissionProfileResponse smart constructor.

Constructors

GetMissionProfileResponse'

Fields

missionProfileId :: Maybe Text

UUID of a mission profile.

missionProfileArn :: Maybe Text

ARN of a mission profile.

trackingConfigArn :: Maybe Text

ARN of a tracking Config.

contactPrePassDurationSeconds :: Maybe Natural

Amount of time prior to contact start you’d like to receive a CloudWatch event indicating an upcoming pass.

contactPostPassDurationSeconds :: Maybe Natural

Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

name :: Maybe Text

Name of a mission profile.

dataflowEdges :: Maybe [NonEmpty Text]

A list of lists of ARNs. Each list of ARNs is an edge, with a from Config and a to Config.

region :: Maybe Text

Region of a mission profile.

minimumViableContactDurationSeconds :: Maybe Natural

Smallest amount of time in seconds that you’d like to see for an available contact. AWS Ground Station will not present you with contacts shorter than this duration.

tags :: Maybe (HashMap Text Text)

Tags assigned to a mission profile.

httpStatus :: Int

The response's http status code.

newGetMissionProfileResponse

Arguments

:: Int
$sel:httpStatus:GetMissionProfileResponse'

-> GetMissionProfileResponse

Create a value of GetMissionProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:missionProfileId:GetMissionProfile', getMissionProfileResponse_missionProfileId - UUID of a mission profile.

$sel:missionProfileArn:GetMissionProfileResponse', getMissionProfileResponse_missionProfileArn - ARN of a mission profile.

$sel:trackingConfigArn:GetMissionProfileResponse', getMissionProfileResponse_trackingConfigArn - ARN of a tracking Config.

$sel:contactPrePassDurationSeconds:GetMissionProfileResponse', getMissionProfileResponse_contactPrePassDurationSeconds - Amount of time prior to contact start you’d like to receive a CloudWatch event indicating an upcoming pass.

$sel:contactPostPassDurationSeconds:GetMissionProfileResponse', getMissionProfileResponse_contactPostPassDurationSeconds - Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

$sel:name:GetMissionProfileResponse', getMissionProfileResponse_name - Name of a mission profile.

$sel:dataflowEdges:GetMissionProfileResponse', getMissionProfileResponse_dataflowEdges - A list of lists of ARNs. Each list of ARNs is an edge, with a from Config and a to Config.

$sel:region:GetMissionProfileResponse', getMissionProfileResponse_region - Region of a mission profile.

$sel:minimumViableContactDurationSeconds:GetMissionProfileResponse', getMissionProfileResponse_minimumViableContactDurationSeconds - Smallest amount of time in seconds that you’d like to see for an available contact. AWS Ground Station will not present you with contacts shorter than this duration.

$sel:tags:GetMissionProfileResponse', getMissionProfileResponse_tags - Tags assigned to a mission profile.

$sel:httpStatus:GetMissionProfileResponse', getMissionProfileResponse_httpStatus - The response's http status code.

Response Lenses

getMissionProfileResponse_missionProfileId :: Lens' GetMissionProfileResponse (Maybe Text)

UUID of a mission profile.

getMissionProfileResponse_missionProfileArn :: Lens' GetMissionProfileResponse (Maybe Text)

ARN of a mission profile.

getMissionProfileResponse_trackingConfigArn :: Lens' GetMissionProfileResponse (Maybe Text)

ARN of a tracking Config.

getMissionProfileResponse_contactPrePassDurationSeconds :: Lens' GetMissionProfileResponse (Maybe Natural)

Amount of time prior to contact start you’d like to receive a CloudWatch event indicating an upcoming pass.

getMissionProfileResponse_contactPostPassDurationSeconds :: Lens' GetMissionProfileResponse (Maybe Natural)

Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

getMissionProfileResponse_name :: Lens' GetMissionProfileResponse (Maybe Text)

Name of a mission profile.

getMissionProfileResponse_dataflowEdges :: Lens' GetMissionProfileResponse (Maybe [NonEmpty Text])

A list of lists of ARNs. Each list of ARNs is an edge, with a from Config and a to Config.

getMissionProfileResponse_region :: Lens' GetMissionProfileResponse (Maybe Text)

Region of a mission profile.

getMissionProfileResponse_minimumViableContactDurationSeconds :: Lens' GetMissionProfileResponse (Maybe Natural)

Smallest amount of time in seconds that you’d like to see for an available contact. AWS Ground Station will not present you with contacts shorter than this duration.

getMissionProfileResponse_tags :: Lens' GetMissionProfileResponse (Maybe (HashMap Text Text))

Tags assigned to a mission profile.

getMissionProfileResponse_httpStatus :: Lens' GetMissionProfileResponse Int

The response's http status code.