Network.AWS.GroundStation.Lens

Description

Operations

DeleteDataflowEndpointGroup

deleteDataflowEndpointGroup_dataflowEndpointGroupId :: Lens' DeleteDataflowEndpointGroup Text

UUID of a dataflow endpoint group.

dataflowEndpointGroupIdResponse_dataflowEndpointGroupId :: Lens' DataflowEndpointGroupIdResponse (Maybe Text)

UUID of a dataflow endpoint group.

ListSatellites

listSatellites_nextToken :: Lens' ListSatellites (Maybe Text)

Next token that can be supplied in the next call to get the next page of satellites.

listSatellites_maxResults :: Lens' ListSatellites (Maybe Int)

Maximum number of satellites returned.

listSatellitesResponse_satellites :: Lens' ListSatellitesResponse (Maybe [SatelliteListItem])

List of satellites.

listSatellitesResponse_nextToken :: Lens' ListSatellitesResponse (Maybe Text)

Next token that can be supplied in the next call to get the next page of satellites.

listSatellitesResponse_httpStatus :: Lens' ListSatellitesResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

ARN of a resource.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

Tags assigned to a resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

GetMinuteUsage

getMinuteUsage_month :: Lens' GetMinuteUsage Int

The month being requested, with a value of 1-12.

getMinuteUsage_year :: Lens' GetMinuteUsage Int

The year being requested, in the format of YYYY.

getMinuteUsageResponse_isReservedMinutesCustomer :: Lens' GetMinuteUsageResponse (Maybe Bool)

Returns whether or not an account has signed up for the reserved minutes pricing plan, specific to the month being requested.

getMinuteUsageResponse_upcomingMinutesScheduled :: Lens' GetMinuteUsageResponse (Maybe Int)

Upcoming minutes scheduled for an account, specific to the month being requested.

getMinuteUsageResponse_totalScheduledMinutes :: Lens' GetMinuteUsageResponse (Maybe Int)

Total scheduled minutes for an account, specific to the month being requested.

getMinuteUsageResponse_totalReservedMinuteAllocation :: Lens' GetMinuteUsageResponse (Maybe Int)

Total number of reserved minutes allocated, specific to the month being requested.

getMinuteUsageResponse_estimatedMinutesRemaining :: Lens' GetMinuteUsageResponse (Maybe Int)

Estimated number of minutes remaining for an account, specific to the month being requested.

getMinuteUsageResponse_httpStatus :: Lens' GetMinuteUsageResponse Int

The response's http status code.

DescribeContact

describeContact_contactId :: Lens' DescribeContact Text

UUID of a contact.

describeContactResponse_contactStatus :: Lens' DescribeContactResponse (Maybe ContactStatus)

Status of a contact.

describeContactResponse_missionProfileArn :: Lens' DescribeContactResponse (Maybe Text)

ARN of a mission profile.

describeContactResponse_startTime :: Lens' DescribeContactResponse (Maybe UTCTime)

Start time of a contact.

describeContactResponse_dataflowList :: Lens' DescribeContactResponse (Maybe [DataflowDetail])

List describing source and destination details for each dataflow edge.

describeContactResponse_satelliteArn :: Lens' DescribeContactResponse (Maybe Text)

ARN of a satellite.

describeContactResponse_maximumElevation :: Lens' DescribeContactResponse (Maybe Elevation)

Maximum elevation angle of a contact.

describeContactResponse_groundStation :: Lens' DescribeContactResponse (Maybe Text)

Ground station for a contact.

describeContactResponse_endTime :: Lens' DescribeContactResponse (Maybe UTCTime)

End time of a contact.

describeContactResponse_contactId :: Lens' DescribeContactResponse (Maybe Text)

UUID of a contact.

describeContactResponse_region :: Lens' DescribeContactResponse (Maybe Text)

Region of a contact.

describeContactResponse_postPassEndTime :: Lens' DescribeContactResponse (Maybe UTCTime)

Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

describeContactResponse_prePassStartTime :: Lens' DescribeContactResponse (Maybe UTCTime)

Amount of time prior to contact start you’d like to receive a CloudWatch event indicating an upcoming pass.

describeContactResponse_errorMessage :: Lens' DescribeContactResponse (Maybe Text)

Error message for a contact.

describeContactResponse_tags :: Lens' DescribeContactResponse (Maybe (HashMap Text Text))

Tags assigned to a contact.

describeContactResponse_httpStatus :: Lens' DescribeContactResponse Int

The response's http status code.

ReserveContact

reserveContact_tags :: Lens' ReserveContact (Maybe (HashMap Text Text))

Tags assigned to a contact.

reserveContact_endTime :: Lens' ReserveContact UTCTime

End time of a contact.

reserveContact_groundStation :: Lens' ReserveContact Text

Name of a ground station.

reserveContact_missionProfileArn :: Lens' ReserveContact Text

ARN of a mission profile.

reserveContact_satelliteArn :: Lens' ReserveContact Text

ARN of a satellite

reserveContact_startTime :: Lens' ReserveContact UTCTime

Start time of a contact.

contactIdResponse_contactId :: Lens' ContactIdResponse (Maybe Text)

UUID of a contact.

CreateMissionProfile

createMissionProfile_contactPrePassDurationSeconds :: Lens' CreateMissionProfile (Maybe Natural)

Amount of time prior to contact start you’d like to receive a CloudWatch event indicating an upcoming pass.

createMissionProfile_contactPostPassDurationSeconds :: Lens' CreateMissionProfile (Maybe Natural)

Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

createMissionProfile_tags :: Lens' CreateMissionProfile (Maybe (HashMap Text Text))

Tags assigned to a mission profile.

createMissionProfile_dataflowEdges :: Lens' CreateMissionProfile [NonEmpty Text]

A list of lists of ARNs. Each list of ARNs is an edge, with a from Config and a to Config.

createMissionProfile_minimumViableContactDurationSeconds :: Lens' CreateMissionProfile Natural

Smallest amount of time in seconds that you’d like to see for an available contact. AWS Ground Station will not present you with contacts shorter than this duration.

createMissionProfile_name :: Lens' CreateMissionProfile Text

Name of a mission profile.

createMissionProfile_trackingConfigArn :: Lens' CreateMissionProfile Text

ARN of a tracking Config.

missionProfileIdResponse_missionProfileId :: Lens' MissionProfileIdResponse (Maybe Text)

UUID of a mission profile.

ListGroundStations

listGroundStations_satelliteId :: Lens' ListGroundStations (Maybe Text)

Satellite ID to retrieve on-boarded ground stations.

listGroundStations_nextToken :: Lens' ListGroundStations (Maybe Text)

Next token that can be supplied in the next call to get the next page of ground stations.

listGroundStations_maxResults :: Lens' ListGroundStations (Maybe Int)

Maximum number of ground stations returned.

listGroundStationsResponse_nextToken :: Lens' ListGroundStationsResponse (Maybe Text)

Next token that can be supplied in the next call to get the next page of ground stations.

listGroundStationsResponse_groundStationList :: Lens' ListGroundStationsResponse (Maybe [GroundStationData])

List of ground stations.

listGroundStationsResponse_httpStatus :: Lens' ListGroundStationsResponse Int

The response's http status code.

CreateConfig

createConfig_tags :: Lens' CreateConfig (Maybe (HashMap Text Text))

Tags assigned to a Config.

createConfig_configData :: Lens' CreateConfig ConfigTypeData

Parameters of a Config.

createConfig_name :: Lens' CreateConfig Text

Name of a Config.

configIdResponse_configArn :: Lens' ConfigIdResponse (Maybe Text)

ARN of a Config.

configIdResponse_configId :: Lens' ConfigIdResponse (Maybe Text)

UUID of a Config.

configIdResponse_configType :: Lens' ConfigIdResponse (Maybe ConfigCapabilityType)

Type of a Config.

ListMissionProfiles

listMissionProfiles_nextToken :: Lens' ListMissionProfiles (Maybe Text)

Next token returned in the request of a previous ListMissionProfiles call. Used to get the next page of results.

listMissionProfiles_maxResults :: Lens' ListMissionProfiles (Maybe Int)

Maximum number of mission profiles returned.

listMissionProfilesResponse_nextToken :: Lens' ListMissionProfilesResponse (Maybe Text)

Next token returned in the response of a previous ListMissionProfiles call. Used to get the next page of results.

listMissionProfilesResponse_missionProfileList :: Lens' ListMissionProfilesResponse (Maybe [MissionProfileListItem])

List of mission profiles.

listMissionProfilesResponse_httpStatus :: Lens' ListMissionProfilesResponse Int

The response's http status code.

GetMissionProfile

getMissionProfile_missionProfileId :: Lens' GetMissionProfile Text

UUID of a mission profile.

getMissionProfileResponse_missionProfileId :: Lens' GetMissionProfileResponse (Maybe Text)

UUID of a mission profile.

getMissionProfileResponse_missionProfileArn :: Lens' GetMissionProfileResponse (Maybe Text)

ARN of a mission profile.

getMissionProfileResponse_trackingConfigArn :: Lens' GetMissionProfileResponse (Maybe Text)

ARN of a tracking Config.

getMissionProfileResponse_contactPrePassDurationSeconds :: Lens' GetMissionProfileResponse (Maybe Natural)

Amount of time prior to contact start you’d like to receive a CloudWatch event indicating an upcoming pass.

getMissionProfileResponse_contactPostPassDurationSeconds :: Lens' GetMissionProfileResponse (Maybe Natural)

Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

getMissionProfileResponse_name :: Lens' GetMissionProfileResponse (Maybe Text)

Name of a mission profile.

getMissionProfileResponse_dataflowEdges :: Lens' GetMissionProfileResponse (Maybe [NonEmpty Text])

A list of lists of ARNs. Each list of ARNs is an edge, with a from Config and a to Config.

getMissionProfileResponse_region :: Lens' GetMissionProfileResponse (Maybe Text)

Region of a mission profile.

getMissionProfileResponse_minimumViableContactDurationSeconds :: Lens' GetMissionProfileResponse (Maybe Natural)

Smallest amount of time in seconds that you’d like to see for an available contact. AWS Ground Station will not present you with contacts shorter than this duration.

getMissionProfileResponse_tags :: Lens' GetMissionProfileResponse (Maybe (HashMap Text Text))

Tags assigned to a mission profile.

getMissionProfileResponse_httpStatus :: Lens' GetMissionProfileResponse Int

The response's http status code.

GetConfig

getConfig_configId :: Lens' GetConfig Text

UUID of a Config.

getConfig_configType :: Lens' GetConfig ConfigCapabilityType

Type of a Config.

getConfigResponse_configType :: Lens' GetConfigResponse (Maybe ConfigCapabilityType)

Type of a Config.

getConfigResponse_tags :: Lens' GetConfigResponse (Maybe (HashMap Text Text))

Tags assigned to a Config.

getConfigResponse_httpStatus :: Lens' GetConfigResponse Int

The response's http status code.

getConfigResponse_configArn :: Lens' GetConfigResponse Text

ARN of a Config

getConfigResponse_configData :: Lens' GetConfigResponse ConfigTypeData

Data elements in a Config.

getConfigResponse_configId :: Lens' GetConfigResponse Text

UUID of a Config.

getConfigResponse_name :: Lens' GetConfigResponse Text

Name of a Config.

ListDataflowEndpointGroups

listDataflowEndpointGroups_nextToken :: Lens' ListDataflowEndpointGroups (Maybe Text)

Next token returned in the request of a previous ListDataflowEndpointGroups call. Used to get the next page of results.

listDataflowEndpointGroups_maxResults :: Lens' ListDataflowEndpointGroups (Maybe Int)

Maximum number of dataflow endpoint groups returned.

listDataflowEndpointGroupsResponse_dataflowEndpointGroupList :: Lens' ListDataflowEndpointGroupsResponse (Maybe [DataflowEndpointListItem])

A list of dataflow endpoint groups.

listDataflowEndpointGroupsResponse_nextToken :: Lens' ListDataflowEndpointGroupsResponse (Maybe Text)

Next token returned in the response of a previous ListDataflowEndpointGroups call. Used to get the next page of results.

listDataflowEndpointGroupsResponse_httpStatus :: Lens' ListDataflowEndpointGroupsResponse Int

The response's http status code.

CreateDataflowEndpointGroup

createDataflowEndpointGroup_tags :: Lens' CreateDataflowEndpointGroup (Maybe (HashMap Text Text))

Tags of a dataflow endpoint group.

createDataflowEndpointGroup_endpointDetails :: Lens' CreateDataflowEndpointGroup [EndpointDetails]

Endpoint details of each endpoint in the dataflow endpoint group.

dataflowEndpointGroupIdResponse_dataflowEndpointGroupId :: Lens' DataflowEndpointGroupIdResponse (Maybe Text)

UUID of a dataflow endpoint group.

GetSatellite

getSatellite_satelliteId :: Lens' GetSatellite Text

UUID of a satellite.

getSatelliteResponse_satelliteId :: Lens' GetSatelliteResponse (Maybe Text)

UUID of a satellite.

getSatelliteResponse_satelliteArn :: Lens' GetSatelliteResponse (Maybe Text)

ARN of a satellite.

getSatelliteResponse_groundStations :: Lens' GetSatelliteResponse (Maybe [Text])

A list of ground stations to which the satellite is on-boarded.

getSatelliteResponse_noradSatelliteID :: Lens' GetSatelliteResponse (Maybe Natural)

NORAD satellite ID number.

getSatelliteResponse_httpStatus :: Lens' GetSatelliteResponse Int

The response's http status code.

GetDataflowEndpointGroup

getDataflowEndpointGroup_dataflowEndpointGroupId :: Lens' GetDataflowEndpointGroup Text

UUID of a dataflow endpoint group.

getDataflowEndpointGroupResponse_endpointsDetails :: Lens' GetDataflowEndpointGroupResponse (Maybe [EndpointDetails])

Details of a dataflow endpoint.

getDataflowEndpointGroupResponse_dataflowEndpointGroupArn :: Lens' GetDataflowEndpointGroupResponse (Maybe Text)

ARN of a dataflow endpoint group.

getDataflowEndpointGroupResponse_dataflowEndpointGroupId :: Lens' GetDataflowEndpointGroupResponse (Maybe Text)

UUID of a dataflow endpoint group.

getDataflowEndpointGroupResponse_tags :: Lens' GetDataflowEndpointGroupResponse (Maybe (HashMap Text Text))

Tags assigned to a dataflow endpoint group.

getDataflowEndpointGroupResponse_httpStatus :: Lens' GetDataflowEndpointGroupResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

ARN of a resource tag.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

Tags assigned to a resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

ListConfigs

listConfigs_nextToken :: Lens' ListConfigs (Maybe Text)

Next token returned in the request of a previous ListConfigs call. Used to get the next page of results.

listConfigs_maxResults :: Lens' ListConfigs (Maybe Int)

Maximum number of Configs returned.

listConfigsResponse_configList :: Lens' ListConfigsResponse (Maybe [ConfigListItem])

List of Config items.

listConfigsResponse_nextToken :: Lens' ListConfigsResponse (Maybe Text)

Next token returned in the response of a previous ListConfigs call. Used to get the next page of results.

listConfigsResponse_httpStatus :: Lens' ListConfigsResponse Int

The response's http status code.

UpdateConfig

updateConfig_configData :: Lens' UpdateConfig ConfigTypeData

Parameters of a Config.

updateConfig_configId :: Lens' UpdateConfig Text

UUID of a Config.

updateConfig_configType :: Lens' UpdateConfig ConfigCapabilityType

Type of a Config.

updateConfig_name :: Lens' UpdateConfig Text

Name of a Config.

configIdResponse_configArn :: Lens' ConfigIdResponse (Maybe Text)

ARN of a Config.

configIdResponse_configId :: Lens' ConfigIdResponse (Maybe Text)

UUID of a Config.

configIdResponse_configType :: Lens' ConfigIdResponse (Maybe ConfigCapabilityType)

Type of a Config.

DeleteConfig

deleteConfig_configId :: Lens' DeleteConfig Text

UUID of a Config.

deleteConfig_configType :: Lens' DeleteConfig ConfigCapabilityType

Type of a Config.

configIdResponse_configArn :: Lens' ConfigIdResponse (Maybe Text)

ARN of a Config.

configIdResponse_configId :: Lens' ConfigIdResponse (Maybe Text)

UUID of a Config.

configIdResponse_configType :: Lens' ConfigIdResponse (Maybe ConfigCapabilityType)

Type of a Config.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

ARN of a resource.

untagResource_tagKeys :: Lens' UntagResource [Text]

Keys of a resource tag.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

UpdateMissionProfile

updateMissionProfile_trackingConfigArn :: Lens' UpdateMissionProfile (Maybe Text)

ARN of a tracking Config.

updateMissionProfile_contactPrePassDurationSeconds :: Lens' UpdateMissionProfile (Maybe Natural)

Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

updateMissionProfile_contactPostPassDurationSeconds :: Lens' UpdateMissionProfile (Maybe Natural)

Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

updateMissionProfile_name :: Lens' UpdateMissionProfile (Maybe Text)

Name of a mission profile.

updateMissionProfile_dataflowEdges :: Lens' UpdateMissionProfile (Maybe [NonEmpty Text])

A list of lists of ARNs. Each list of ARNs is an edge, with a from Config and a to Config.

updateMissionProfile_minimumViableContactDurationSeconds :: Lens' UpdateMissionProfile (Maybe Natural)

Smallest amount of time in seconds that you’d like to see for an available contact. AWS Ground Station will not present you with contacts shorter than this duration.

updateMissionProfile_missionProfileId :: Lens' UpdateMissionProfile Text

UUID of a mission profile.

missionProfileIdResponse_missionProfileId :: Lens' MissionProfileIdResponse (Maybe Text)

UUID of a mission profile.

DeleteMissionProfile

deleteMissionProfile_missionProfileId :: Lens' DeleteMissionProfile Text

UUID of a mission profile.

missionProfileIdResponse_missionProfileId :: Lens' MissionProfileIdResponse (Maybe Text)

UUID of a mission profile.

CancelContact

cancelContact_contactId :: Lens' CancelContact Text

UUID of a contact.

contactIdResponse_contactId :: Lens' ContactIdResponse (Maybe Text)

UUID of a contact.

ListContacts

listContacts_missionProfileArn :: Lens' ListContacts (Maybe Text)

ARN of a mission profile.

listContacts_satelliteArn :: Lens' ListContacts (Maybe Text)

ARN of a satellite.

listContacts_nextToken :: Lens' ListContacts (Maybe Text)

Next token returned in the request of a previous ListContacts call. Used to get the next page of results.

listContacts_groundStation :: Lens' ListContacts (Maybe Text)

Name of a ground station.

listContacts_maxResults :: Lens' ListContacts (Maybe Int)

Maximum number of contacts returned.

listContacts_endTime :: Lens' ListContacts UTCTime

End time of a contact.

listContacts_startTime :: Lens' ListContacts UTCTime

Start time of a contact.

listContacts_statusList :: Lens' ListContacts [ContactStatus]

Status of a contact reservation.

listContactsResponse_contactList :: Lens' ListContactsResponse (Maybe [ContactData])

List of contacts.

listContactsResponse_nextToken :: Lens' ListContactsResponse (Maybe Text)

Next token returned in the response of a previous ListContacts call. Used to get the next page of results.

listContactsResponse_httpStatus :: Lens' ListContactsResponse Int

The response's http status code.

Types

AntennaDemodDecodeDetails

antennaDemodDecodeDetails_outputNode :: Lens' AntennaDemodDecodeDetails (Maybe Text)

Name of an antenna demod decode output node used in a contact.

AntennaDownlinkConfig

antennaDownlinkConfig_spectrumConfig :: Lens' AntennaDownlinkConfig SpectrumConfig

Object that describes a spectral Config.

AntennaDownlinkDemodDecodeConfig

antennaDownlinkDemodDecodeConfig_decodeConfig :: Lens' AntennaDownlinkDemodDecodeConfig DecodeConfig

Information about the decode Config.

antennaDownlinkDemodDecodeConfig_demodulationConfig :: Lens' AntennaDownlinkDemodDecodeConfig DemodulationConfig

Information about the demodulation Config.

antennaDownlinkDemodDecodeConfig_spectrumConfig :: Lens' AntennaDownlinkDemodDecodeConfig SpectrumConfig

Information about the spectral Config.

AntennaUplinkConfig

antennaUplinkConfig_transmitDisabled :: Lens' AntennaUplinkConfig (Maybe Bool)

Whether or not uplink transmit is disabled.

antennaUplinkConfig_spectrumConfig :: Lens' AntennaUplinkConfig UplinkSpectrumConfig

Information about the uplink spectral Config.

antennaUplinkConfig_targetEirp :: Lens' AntennaUplinkConfig Eirp

EIRP of the target.

ConfigDetails

configDetails_antennaDemodDecodeDetails :: Lens' ConfigDetails (Maybe AntennaDemodDecodeDetails)

Details for antenna demod decode Config in a contact.

configDetails_s3RecordingDetails :: Lens' ConfigDetails (Maybe S3RecordingDetails)

Details for an S3 recording Config in a contact.

configDetails_endpointDetails :: Lens' ConfigDetails (Maybe EndpointDetails)

Undocumented member.

ConfigIdResponse

configIdResponse_configArn :: Lens' ConfigIdResponse (Maybe Text)

ARN of a Config.

configIdResponse_configId :: Lens' ConfigIdResponse (Maybe Text)

UUID of a Config.

configIdResponse_configType :: Lens' ConfigIdResponse (Maybe ConfigCapabilityType)

Type of a Config.

ConfigListItem

configListItem_configArn :: Lens' ConfigListItem (Maybe Text)

ARN of a Config.

configListItem_name :: Lens' ConfigListItem (Maybe Text)

Name of a Config.

configListItem_configId :: Lens' ConfigListItem (Maybe Text)

UUID of a Config.

configListItem_configType :: Lens' ConfigListItem (Maybe ConfigCapabilityType)

Type of a Config.

ConfigTypeData

configTypeData_antennaDownlinkDemodDecodeConfig :: Lens' ConfigTypeData (Maybe AntennaDownlinkDemodDecodeConfig)

Information about how AWS Ground Station should conﬁgure an antenna for downlink demod decode during a contact.

configTypeData_antennaDownlinkConfig :: Lens' ConfigTypeData (Maybe AntennaDownlinkConfig)

Information about how AWS Ground Station should configure an antenna for downlink during a contact.

configTypeData_antennaUplinkConfig :: Lens' ConfigTypeData (Maybe AntennaUplinkConfig)

Information about how AWS Ground Station should conﬁgure an antenna for uplink during a contact.

configTypeData_uplinkEchoConfig :: Lens' ConfigTypeData (Maybe UplinkEchoConfig)

Information about an uplink echo Config.

Parameters from the AntennaUplinkConfig, corresponding to the specified AntennaUplinkConfigArn, are used when this UplinkEchoConfig is used in a contact.

configTypeData_s3RecordingConfig :: Lens' ConfigTypeData (Maybe S3RecordingConfig)

Information about an S3 recording Config.

configTypeData_trackingConfig :: Lens' ConfigTypeData (Maybe TrackingConfig)

Object that determines whether tracking should be used during a contact executed with this Config in the mission profile.

configTypeData_dataflowEndpointConfig :: Lens' ConfigTypeData (Maybe DataflowEndpointConfig)

Information about the dataflow endpoint Config.

ContactData

contactData_contactStatus :: Lens' ContactData (Maybe ContactStatus)

Status of a contact.

contactData_missionProfileArn :: Lens' ContactData (Maybe Text)

ARN of a mission profile.

contactData_startTime :: Lens' ContactData (Maybe UTCTime)

Start time of a contact.

contactData_satelliteArn :: Lens' ContactData (Maybe Text)

ARN of a satellite.

contactData_maximumElevation :: Lens' ContactData (Maybe Elevation)

Maximum elevation angle of a contact.

contactData_groundStation :: Lens' ContactData (Maybe Text)

Name of a ground station.

contactData_endTime :: Lens' ContactData (Maybe UTCTime)

End time of a contact.

contactData_contactId :: Lens' ContactData (Maybe Text)

UUID of a contact.

contactData_region :: Lens' ContactData (Maybe Text)

Region of a contact.

contactData_postPassEndTime :: Lens' ContactData (Maybe UTCTime)

Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

contactData_prePassStartTime :: Lens' ContactData (Maybe UTCTime)

Amount of time prior to contact start you’d like to receive a CloudWatch event indicating an upcoming pass.

contactData_errorMessage :: Lens' ContactData (Maybe Text)

Error message of a contact.

contactData_tags :: Lens' ContactData (Maybe (HashMap Text Text))

Tags assigned to a contact.

ContactIdResponse

contactIdResponse_contactId :: Lens' ContactIdResponse (Maybe Text)

UUID of a contact.

DataflowDetail

dataflowDetail_destination :: Lens' DataflowDetail (Maybe Destination)

Undocumented member.

dataflowDetail_source :: Lens' DataflowDetail (Maybe Source)

Undocumented member.

dataflowDetail_errorMessage :: Lens' DataflowDetail (Maybe Text)

Error message for a dataflow.

DataflowEndpoint

dataflowEndpoint_mtu :: Lens' DataflowEndpoint (Maybe Natural)

Maximum transmission unit (MTU) size in bytes of a dataflow endpoint.

dataflowEndpoint_status :: Lens' DataflowEndpoint (Maybe EndpointStatus)

Status of a dataflow endpoint.

dataflowEndpoint_address :: Lens' DataflowEndpoint (Maybe SocketAddress)

Socket address of a dataflow endpoint.

dataflowEndpoint_name :: Lens' DataflowEndpoint (Maybe Text)

Name of a dataflow endpoint.

DataflowEndpointConfig

dataflowEndpointConfig_dataflowEndpointRegion :: Lens' DataflowEndpointConfig (Maybe Text)

Region of a dataflow endpoint.

dataflowEndpointConfig_dataflowEndpointName :: Lens' DataflowEndpointConfig Text

Name of a dataflow endpoint.

DataflowEndpointGroupIdResponse

dataflowEndpointGroupIdResponse_dataflowEndpointGroupId :: Lens' DataflowEndpointGroupIdResponse (Maybe Text)

UUID of a dataflow endpoint group.

DataflowEndpointListItem

dataflowEndpointListItem_dataflowEndpointGroupArn :: Lens' DataflowEndpointListItem (Maybe Text)

ARN of a dataflow endpoint group.

dataflowEndpointListItem_dataflowEndpointGroupId :: Lens' DataflowEndpointListItem (Maybe Text)

UUID of a dataflow endpoint group.

DecodeConfig

decodeConfig_unvalidatedJSON :: Lens' DecodeConfig Text

Unvalidated JSON of a decode Config.

DemodulationConfig

demodulationConfig_unvalidatedJSON :: Lens' DemodulationConfig Text

Unvalidated JSON of a demodulation Config.

Destination

destination_dataflowDestinationRegion :: Lens' Destination (Maybe Text)

Region of a dataflow destination.

destination_configId :: Lens' Destination (Maybe Text)

UUID of a Config.

destination_configType :: Lens' Destination (Maybe ConfigCapabilityType)

Type of a Config.

destination_configDetails :: Lens' Destination (Maybe ConfigDetails)

Additional details for a Config, if type is dataflow endpoint or antenna demod decode.

Eirp

eirp_units :: Lens' Eirp EirpUnits

Units of an EIRP.

eirp_value :: Lens' Eirp Double

Value of an EIRP. Valid values are between 20.0 to 50.0 dBW.

Elevation

elevation_unit :: Lens' Elevation AngleUnits

Elevation angle units.

elevation_value :: Lens' Elevation Double

Elevation angle value.

EndpointDetails

endpointDetails_securityDetails :: Lens' EndpointDetails (Maybe SecurityDetails)

Endpoint security details.

endpointDetails_endpoint :: Lens' EndpointDetails (Maybe DataflowEndpoint)

A dataflow endpoint.

Frequency

frequency_units :: Lens' Frequency FrequencyUnits

Frequency units.

frequency_value :: Lens' Frequency Double

Frequency value. Valid values are between 2200 to 2300 MHz and 7750 to 8400 MHz for downlink and 2025 to 2120 MHz for uplink.

FrequencyBandwidth

frequencyBandwidth_units :: Lens' FrequencyBandwidth BandwidthUnits

Frequency bandwidth units.

frequencyBandwidth_value :: Lens' FrequencyBandwidth Double

Frequency bandwidth value. AWS Ground Station currently has the following bandwidth limitations:

For AntennaDownlinkDemodDecodeconfig, valid values are between 125 kHz to 650 MHz.

For AntennaDownlinkconfig, valid values are between 10 kHz to 54 MHz.

For AntennaUplinkConfig, valid values are between 10 kHz to 54 MHz.

GroundStationData

groundStationData_groundStationId :: Lens' GroundStationData (Maybe Text)

UUID of a ground station.

groundStationData_groundStationName :: Lens' GroundStationData (Maybe Text)

Name of a ground station.

groundStationData_region :: Lens' GroundStationData (Maybe Text)

Ground station Region.

MissionProfileIdResponse

missionProfileIdResponse_missionProfileId :: Lens' MissionProfileIdResponse (Maybe Text)

UUID of a mission profile.

MissionProfileListItem

missionProfileListItem_missionProfileId :: Lens' MissionProfileListItem (Maybe Text)

UUID of a mission profile.

missionProfileListItem_missionProfileArn :: Lens' MissionProfileListItem (Maybe Text)

ARN of a mission profile.

missionProfileListItem_name :: Lens' MissionProfileListItem (Maybe Text)

Name of a mission profile.

missionProfileListItem_region :: Lens' MissionProfileListItem (Maybe Text)

Region of a mission profile.

S3RecordingConfig

s3RecordingConfig_prefix :: Lens' S3RecordingConfig (Maybe Text)

S3 Key prefix to prefice data files.

s3RecordingConfig_bucketArn :: Lens' S3RecordingConfig Text

ARN of the bucket to record to.

s3RecordingConfig_roleArn :: Lens' S3RecordingConfig Text

ARN of the role Ground Station assumes to write data to the bucket.

S3RecordingDetails

s3RecordingDetails_keyTemplate :: Lens' S3RecordingDetails (Maybe Text)

Template of the S3 key used.

s3RecordingDetails_bucketArn :: Lens' S3RecordingDetails (Maybe Text)

ARN of the bucket used.

SatelliteListItem

satelliteListItem_satelliteId :: Lens' SatelliteListItem (Maybe Text)

UUID of a satellite.

satelliteListItem_satelliteArn :: Lens' SatelliteListItem (Maybe Text)

ARN of a satellite.

satelliteListItem_groundStations :: Lens' SatelliteListItem (Maybe [Text])

A list of ground stations to which the satellite is on-boarded.

satelliteListItem_noradSatelliteID :: Lens' SatelliteListItem (Maybe Natural)

NORAD satellite ID number.

SecurityDetails

securityDetails_roleArn :: Lens' SecurityDetails Text

ARN to a role needed for connecting streams to your instances.

securityDetails_securityGroupIds :: Lens' SecurityDetails [Text]

The security groups to attach to the elastic network interfaces.

securityDetails_subnetIds :: Lens' SecurityDetails [Text]

A list of subnets where AWS Ground Station places elastic network interfaces to send streams to your instances.

SocketAddress

socketAddress_name :: Lens' SocketAddress Text

Name of a socket address.

socketAddress_port :: Lens' SocketAddress Int

Port of a socket address.

Source

source_dataflowSourceRegion :: Lens' Source (Maybe Text)

Region of a dataflow source.

source_configId :: Lens' Source (Maybe Text)

UUID of a Config.

source_configType :: Lens' Source (Maybe ConfigCapabilityType)

Type of a Config.

source_configDetails :: Lens' Source (Maybe ConfigDetails)

Additional details for a Config, if type is dataflow endpoint or antenna demod decode.

SpectrumConfig

spectrumConfig_polarization :: Lens' SpectrumConfig (Maybe Polarization)

Polarization of a spectral Config. Capturing both "RIGHT_HAND" and "LEFT_HAND" polarization requires two separate configs.

spectrumConfig_bandwidth :: Lens' SpectrumConfig FrequencyBandwidth

Bandwidth of a spectral Config. AWS Ground Station currently has the following bandwidth limitations:

For AntennaDownlinkDemodDecodeconfig, valid values are between 125 kHz to 650 MHz.

For AntennaDownlinkconfig valid values are between 10 kHz to 54 MHz.

For AntennaUplinkConfig, valid values are between 10 kHz to 54 MHz.

spectrumConfig_centerFrequency :: Lens' SpectrumConfig Frequency

Center frequency of a spectral Config. Valid values are between 2200 to 2300 MHz and 7750 to 8400 MHz for downlink and 2025 to 2120 MHz for uplink.

TrackingConfig

trackingConfig_autotrack :: Lens' TrackingConfig Criticality

Current setting for autotrack.

UplinkEchoConfig

uplinkEchoConfig_antennaUplinkConfigArn :: Lens' UplinkEchoConfig Text

ARN of an uplink Config.

uplinkEchoConfig_enabled :: Lens' UplinkEchoConfig Bool

Whether or not an uplink Config is enabled.

UplinkSpectrumConfig

uplinkSpectrumConfig_polarization :: Lens' UplinkSpectrumConfig (Maybe Polarization)

Polarization of an uplink spectral Config. Capturing both "RIGHT_HAND" and "LEFT_HAND" polarization requires two separate configs.

uplinkSpectrumConfig_centerFrequency :: Lens' UplinkSpectrumConfig Frequency

Center frequency of an uplink spectral Config. Valid values are between 2025 to 2120 MHz.