Network.AWS.GroundStation.ListConfigs

Description

Returns a list of Config objects.

This operation returns paginated results.

Creating a Request

data ListConfigs

See: newListConfigs smart constructor.

Constructors

ListConfigs'

Fields

nextToken :: Maybe Text

Next token returned in the request of a previous ListConfigs call. Used to get the next page of results.

maxResults :: Maybe Int

Maximum number of Configs returned.

newListConfigs :: ListConfigs

Create a value of ListConfigs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListConfigs', listConfigs_nextToken - Next token returned in the request of a previous ListConfigs call. Used to get the next page of results.

$sel:maxResults:ListConfigs', listConfigs_maxResults - Maximum number of Configs returned.

Request Lenses

listConfigs_nextToken :: Lens' ListConfigs (Maybe Text)

Next token returned in the request of a previous ListConfigs call. Used to get the next page of results.

listConfigs_maxResults :: Lens' ListConfigs (Maybe Int)

Maximum number of Configs returned.

Destructuring the Response

data ListConfigsResponse

See: newListConfigsResponse smart constructor.

Constructors

ListConfigsResponse'

Fields

configList :: Maybe [ConfigListItem]

List of Config items.

nextToken :: Maybe Text

Next token returned in the response of a previous ListConfigs call. Used to get the next page of results.

httpStatus :: Int

The response's http status code.

newListConfigsResponse

Arguments

:: Int
$sel:httpStatus:ListConfigsResponse'

-> ListConfigsResponse

Create a value of ListConfigsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configList:ListConfigsResponse', listConfigsResponse_configList - List of Config items.

$sel:nextToken:ListConfigs', listConfigsResponse_nextToken - Next token returned in the response of a previous ListConfigs call. Used to get the next page of results.

$sel:httpStatus:ListConfigsResponse', listConfigsResponse_httpStatus - The response's http status code.

Response Lenses

listConfigsResponse_configList :: Lens' ListConfigsResponse (Maybe [ConfigListItem])

List of Config items.

listConfigsResponse_nextToken :: Lens' ListConfigsResponse (Maybe Text)

Next token returned in the response of a previous ListConfigs call. Used to get the next page of results.

listConfigsResponse_httpStatus :: Lens' ListConfigsResponse Int

The response's http status code.