Network.AWS.GroundStation.ListContacts

Description

Returns a list of contacts.

If statusList contains AVAILABLE, the request must include groundStation, missionprofileArn, and satelliteArn.

This operation returns paginated results.

Creating a Request

data ListContacts

See: newListContacts smart constructor.

Constructors

ListContacts'

Fields

missionProfileArn :: Maybe Text

ARN of a mission profile.

satelliteArn :: Maybe Text

ARN of a satellite.

nextToken :: Maybe Text

Next token returned in the request of a previous ListContacts call. Used to get the next page of results.

groundStation :: Maybe Text

Name of a ground station.

maxResults :: Maybe Int

Maximum number of contacts returned.

endTime :: POSIX

End time of a contact.

startTime :: POSIX

Start time of a contact.

statusList :: [ContactStatus]

Status of a contact reservation.

newListContacts

Arguments

:: UTCTime
$sel:endTime:ListContacts'

-> UTCTime
$sel:startTime:ListContacts'

-> ListContacts

Create a value of ListContacts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:missionProfileArn:ListContacts', listContacts_missionProfileArn - ARN of a mission profile.

$sel:satelliteArn:ListContacts', listContacts_satelliteArn - ARN of a satellite.

$sel:nextToken:ListContacts', listContacts_nextToken - Next token returned in the request of a previous ListContacts call. Used to get the next page of results.

$sel:groundStation:ListContacts', listContacts_groundStation - Name of a ground station.

$sel:maxResults:ListContacts', listContacts_maxResults - Maximum number of contacts returned.

$sel:endTime:ListContacts', listContacts_endTime - End time of a contact.

$sel:startTime:ListContacts', listContacts_startTime - Start time of a contact.

$sel:statusList:ListContacts', listContacts_statusList - Status of a contact reservation.

Request Lenses

listContacts_missionProfileArn :: Lens' ListContacts (Maybe Text)

ARN of a mission profile.

listContacts_satelliteArn :: Lens' ListContacts (Maybe Text)

ARN of a satellite.

listContacts_nextToken :: Lens' ListContacts (Maybe Text)

Next token returned in the request of a previous ListContacts call. Used to get the next page of results.

listContacts_groundStation :: Lens' ListContacts (Maybe Text)

Name of a ground station.

listContacts_maxResults :: Lens' ListContacts (Maybe Int)

Maximum number of contacts returned.

listContacts_endTime :: Lens' ListContacts UTCTime

End time of a contact.

listContacts_startTime :: Lens' ListContacts UTCTime

Start time of a contact.

listContacts_statusList :: Lens' ListContacts [ContactStatus]

Status of a contact reservation.

Destructuring the Response

data ListContactsResponse

See: newListContactsResponse smart constructor.

Constructors

ListContactsResponse'

Fields

contactList :: Maybe [ContactData]

List of contacts.

nextToken :: Maybe Text

Next token returned in the response of a previous ListContacts call. Used to get the next page of results.

httpStatus :: Int

The response's http status code.

newListContactsResponse

Arguments

:: Int
$sel:httpStatus:ListContactsResponse'

-> ListContactsResponse

Create a value of ListContactsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contactList:ListContactsResponse', listContactsResponse_contactList - List of contacts.

$sel:nextToken:ListContacts', listContactsResponse_nextToken - Next token returned in the response of a previous ListContacts call. Used to get the next page of results.

$sel:httpStatus:ListContactsResponse', listContactsResponse_httpStatus - The response's http status code.

Response Lenses

listContactsResponse_contactList :: Lens' ListContactsResponse (Maybe [ContactData])

List of contacts.

listContactsResponse_nextToken :: Lens' ListContactsResponse (Maybe Text)

Next token returned in the response of a previous ListContacts call. Used to get the next page of results.

listContactsResponse_httpStatus :: Lens' ListContactsResponse Int

The response's http status code.