Network.AWS.GroundStation.ReserveContact

Description

Reserves a contact using specified parameters.

Creating a Request

data ReserveContact

See: newReserveContact smart constructor.

Constructors

ReserveContact'

Fields

tags :: Maybe (HashMap Text Text)

Tags assigned to a contact.

endTime :: POSIX

End time of a contact.

groundStation :: Text

Name of a ground station.

missionProfileArn :: Text

ARN of a mission profile.

satelliteArn :: Text

ARN of a satellite

startTime :: POSIX

Start time of a contact.

newReserveContact

Arguments

:: UTCTime
$sel:endTime:ReserveContact'

-> Text
$sel:groundStation:ReserveContact'

-> Text
$sel:missionProfileArn:ReserveContact'

-> Text
$sel:satelliteArn:ReserveContact'

-> UTCTime
$sel:startTime:ReserveContact'

-> ReserveContact

Create a value of ReserveContact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ReserveContact', reserveContact_tags - Tags assigned to a contact.

$sel:endTime:ReserveContact', reserveContact_endTime - End time of a contact.

$sel:groundStation:ReserveContact', reserveContact_groundStation - Name of a ground station.

$sel:missionProfileArn:ReserveContact', reserveContact_missionProfileArn - ARN of a mission profile.

$sel:satelliteArn:ReserveContact', reserveContact_satelliteArn - ARN of a satellite

$sel:startTime:ReserveContact', reserveContact_startTime - Start time of a contact.

Request Lenses

reserveContact_tags :: Lens' ReserveContact (Maybe (HashMap Text Text))

Tags assigned to a contact.

reserveContact_endTime :: Lens' ReserveContact UTCTime

End time of a contact.

reserveContact_groundStation :: Lens' ReserveContact Text

Name of a ground station.

reserveContact_missionProfileArn :: Lens' ReserveContact Text

ARN of a mission profile.

reserveContact_satelliteArn :: Lens' ReserveContact Text

ARN of a satellite

reserveContact_startTime :: Lens' ReserveContact UTCTime

Start time of a contact.

Destructuring the Response

data ContactIdResponse

See: newContactIdResponse smart constructor.

Constructors

ContactIdResponse'

Fields

contactId :: Maybe Text

UUID of a contact.

newContactIdResponse :: ContactIdResponse

Create a value of ContactIdResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contactId:ContactIdResponse', contactIdResponse_contactId - UUID of a contact.

Response Lenses

contactIdResponse_contactId :: Lens' ContactIdResponse (Maybe Text)

UUID of a contact.