Network.AWS.GroundStation.Types.AntennaDownlinkDemodDecodeConfig

Description

Documentation

data AntennaDownlinkDemodDecodeConfig

Information about how AWS Ground Station should conﬁgure an antenna for downlink demod decode during a contact.

See: newAntennaDownlinkDemodDecodeConfig smart constructor.

Constructors

AntennaDownlinkDemodDecodeConfig'

Fields

decodeConfig :: DecodeConfig

Information about the decode Config.

demodulationConfig :: DemodulationConfig

Information about the demodulation Config.

spectrumConfig :: SpectrumConfig

Information about the spectral Config.

newAntennaDownlinkDemodDecodeConfig

Arguments

:: DecodeConfig
$sel:decodeConfig:AntennaDownlinkDemodDecodeConfig'

-> DemodulationConfig
$sel:demodulationConfig:AntennaDownlinkDemodDecodeConfig'

-> SpectrumConfig
$sel:spectrumConfig:AntennaDownlinkDemodDecodeConfig'

-> AntennaDownlinkDemodDecodeConfig

Create a value of AntennaDownlinkDemodDecodeConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:decodeConfig:AntennaDownlinkDemodDecodeConfig', antennaDownlinkDemodDecodeConfig_decodeConfig - Information about the decode Config.

$sel:demodulationConfig:AntennaDownlinkDemodDecodeConfig', antennaDownlinkDemodDecodeConfig_demodulationConfig - Information about the demodulation Config.

$sel:spectrumConfig:AntennaDownlinkDemodDecodeConfig', antennaDownlinkDemodDecodeConfig_spectrumConfig - Information about the spectral Config.

antennaDownlinkDemodDecodeConfig_decodeConfig :: Lens' AntennaDownlinkDemodDecodeConfig DecodeConfig

Information about the decode Config.

antennaDownlinkDemodDecodeConfig_demodulationConfig :: Lens' AntennaDownlinkDemodDecodeConfig DemodulationConfig

Information about the demodulation Config.

antennaDownlinkDemodDecodeConfig_spectrumConfig :: Lens' AntennaDownlinkDemodDecodeConfig SpectrumConfig

Information about the spectral Config.