Network.AWS.GroundStation.Types.ContactData

Description

Documentation

data ContactData

Data describing a contact.

See: newContactData smart constructor.

Constructors

ContactData'

Fields

contactStatus :: Maybe ContactStatus

Status of a contact.

missionProfileArn :: Maybe Text

ARN of a mission profile.

startTime :: Maybe POSIX

Start time of a contact.

satelliteArn :: Maybe Text

ARN of a satellite.

maximumElevation :: Maybe Elevation

Maximum elevation angle of a contact.

groundStation :: Maybe Text

Name of a ground station.

endTime :: Maybe POSIX

End time of a contact.

contactId :: Maybe Text

UUID of a contact.

region :: Maybe Text

Region of a contact.

postPassEndTime :: Maybe POSIX

Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

prePassStartTime :: Maybe POSIX

Amount of time prior to contact start you’d like to receive a CloudWatch event indicating an upcoming pass.

errorMessage :: Maybe Text

Error message of a contact.

tags :: Maybe (HashMap Text Text)

Tags assigned to a contact.

newContactData :: ContactData

Create a value of ContactData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contactStatus:ContactData', contactData_contactStatus - Status of a contact.

$sel:missionProfileArn:ContactData', contactData_missionProfileArn - ARN of a mission profile.

$sel:startTime:ContactData', contactData_startTime - Start time of a contact.

$sel:satelliteArn:ContactData', contactData_satelliteArn - ARN of a satellite.

$sel:maximumElevation:ContactData', contactData_maximumElevation - Maximum elevation angle of a contact.

$sel:groundStation:ContactData', contactData_groundStation - Name of a ground station.

$sel:endTime:ContactData', contactData_endTime - End time of a contact.

$sel:contactId:ContactData', contactData_contactId - UUID of a contact.

$sel:region:ContactData', contactData_region - Region of a contact.

$sel:postPassEndTime:ContactData', contactData_postPassEndTime - Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

$sel:prePassStartTime:ContactData', contactData_prePassStartTime - Amount of time prior to contact start you’d like to receive a CloudWatch event indicating an upcoming pass.

$sel:errorMessage:ContactData', contactData_errorMessage - Error message of a contact.

$sel:tags:ContactData', contactData_tags - Tags assigned to a contact.

contactData_contactStatus :: Lens' ContactData (Maybe ContactStatus)

Status of a contact.

contactData_missionProfileArn :: Lens' ContactData (Maybe Text)

ARN of a mission profile.

contactData_startTime :: Lens' ContactData (Maybe UTCTime)

Start time of a contact.

contactData_satelliteArn :: Lens' ContactData (Maybe Text)

ARN of a satellite.

contactData_maximumElevation :: Lens' ContactData (Maybe Elevation)

Maximum elevation angle of a contact.

contactData_groundStation :: Lens' ContactData (Maybe Text)

Name of a ground station.

contactData_endTime :: Lens' ContactData (Maybe UTCTime)

End time of a contact.

contactData_contactId :: Lens' ContactData (Maybe Text)

UUID of a contact.

contactData_region :: Lens' ContactData (Maybe Text)

Region of a contact.

contactData_postPassEndTime :: Lens' ContactData (Maybe UTCTime)

Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

contactData_prePassStartTime :: Lens' ContactData (Maybe UTCTime)

Amount of time prior to contact start you’d like to receive a CloudWatch event indicating an upcoming pass.

contactData_errorMessage :: Lens' ContactData (Maybe Text)

Error message of a contact.

contactData_tags :: Lens' ContactData (Maybe (HashMap Text Text))

Tags assigned to a contact.