Network.AWS.GroundStation.UpdateMissionProfile

Description

Updates a mission profile.

Updating a mission profile will not update the execution parameters for existing future contacts.

Creating a Request

data UpdateMissionProfile

See: newUpdateMissionProfile smart constructor.

Constructors

UpdateMissionProfile'

Fields

trackingConfigArn :: Maybe Text

ARN of a tracking Config.

contactPrePassDurationSeconds :: Maybe Natural

Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

contactPostPassDurationSeconds :: Maybe Natural

Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

name :: Maybe Text

Name of a mission profile.

dataflowEdges :: Maybe [NonEmpty Text]

A list of lists of ARNs. Each list of ARNs is an edge, with a from Config and a to Config.

minimumViableContactDurationSeconds :: Maybe Natural

Smallest amount of time in seconds that you’d like to see for an available contact. AWS Ground Station will not present you with contacts shorter than this duration.

missionProfileId :: Text

UUID of a mission profile.

newUpdateMissionProfile

Arguments

:: Text
$sel:missionProfileId:UpdateMissionProfile'

-> UpdateMissionProfile

Create a value of UpdateMissionProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trackingConfigArn:UpdateMissionProfile', updateMissionProfile_trackingConfigArn - ARN of a tracking Config.

$sel:contactPrePassDurationSeconds:UpdateMissionProfile', updateMissionProfile_contactPrePassDurationSeconds - Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

$sel:contactPostPassDurationSeconds:UpdateMissionProfile', updateMissionProfile_contactPostPassDurationSeconds - Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

$sel:name:UpdateMissionProfile', updateMissionProfile_name - Name of a mission profile.

$sel:dataflowEdges:UpdateMissionProfile', updateMissionProfile_dataflowEdges - A list of lists of ARNs. Each list of ARNs is an edge, with a from Config and a to Config.

$sel:minimumViableContactDurationSeconds:UpdateMissionProfile', updateMissionProfile_minimumViableContactDurationSeconds - Smallest amount of time in seconds that you’d like to see for an available contact. AWS Ground Station will not present you with contacts shorter than this duration.

$sel:missionProfileId:UpdateMissionProfile', updateMissionProfile_missionProfileId - UUID of a mission profile.

Request Lenses

updateMissionProfile_trackingConfigArn :: Lens' UpdateMissionProfile (Maybe Text)

ARN of a tracking Config.

updateMissionProfile_contactPrePassDurationSeconds :: Lens' UpdateMissionProfile (Maybe Natural)

Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

updateMissionProfile_contactPostPassDurationSeconds :: Lens' UpdateMissionProfile (Maybe Natural)

Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

updateMissionProfile_name :: Lens' UpdateMissionProfile (Maybe Text)

Name of a mission profile.

updateMissionProfile_dataflowEdges :: Lens' UpdateMissionProfile (Maybe [NonEmpty Text])

A list of lists of ARNs. Each list of ARNs is an edge, with a from Config and a to Config.

updateMissionProfile_minimumViableContactDurationSeconds :: Lens' UpdateMissionProfile (Maybe Natural)

Smallest amount of time in seconds that you’d like to see for an available contact. AWS Ground Station will not present you with contacts shorter than this duration.

updateMissionProfile_missionProfileId :: Lens' UpdateMissionProfile Text

UUID of a mission profile.

Destructuring the Response

data MissionProfileIdResponse

See: newMissionProfileIdResponse smart constructor.

Constructors

MissionProfileIdResponse'

Fields

missionProfileId :: Maybe Text

UUID of a mission profile.

newMissionProfileIdResponse :: MissionProfileIdResponse

Create a value of MissionProfileIdResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:missionProfileId:MissionProfileIdResponse', missionProfileIdResponse_missionProfileId - UUID of a mission profile.

Response Lenses

missionProfileIdResponse_missionProfileId :: Lens' MissionProfileIdResponse (Maybe Text)

UUID of a mission profile.