Network.AWS.GuardDuty.CreateIPSet

Description

Creates a new IPSet, which is called a trusted IP list in the console user interface. An IPSet is a list of IP addresses that are trusted for secure communication with AWS infrastructure and applications. GuardDuty doesn't generate findings for IP addresses that are included in IPSets. Only users from the administrator account can use this operation.

Creating a Request

data CreateIPSet

See: newCreateIPSet smart constructor.

Constructors

CreateIPSet'

Fields

clientToken :: Maybe Text

The idempotency token for the create request.

tags :: Maybe (HashMap Text Text)

The tags to be added to a new IP set resource.

detectorId :: Text

The unique ID of the detector of the GuardDuty account that you want to create an IPSet for.

name :: Text

The user-friendly name to identify the IPSet.

Allowed characters are alphanumerics, spaces, hyphens (-), and underscores (_).

format :: IpSetFormat

The format of the file that contains the IPSet.

location :: Text

The URI of the file that contains the IPSet. For example: https://s3.us-west-2.amazonaws.com/my-bucket/my-object-key.

activate :: Bool

A Boolean value that indicates whether GuardDuty is to start using the uploaded IPSet.

newCreateIPSet

Arguments

:: Text
$sel:detectorId:CreateIPSet'

-> Text
$sel:name:CreateIPSet'

-> IpSetFormat
$sel:format:CreateIPSet'

-> Text
$sel:location:CreateIPSet'

-> Bool
$sel:activate:CreateIPSet'

-> CreateIPSet

Create a value of CreateIPSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateIPSet', createIPSet_clientToken - The idempotency token for the create request.

$sel:tags:CreateIPSet', createIPSet_tags - The tags to be added to a new IP set resource.

$sel:detectorId:CreateIPSet', createIPSet_detectorId - The unique ID of the detector of the GuardDuty account that you want to create an IPSet for.

$sel:name:CreateIPSet', createIPSet_name - The user-friendly name to identify the IPSet.

Allowed characters are alphanumerics, spaces, hyphens (-), and underscores (_).

$sel:format:CreateIPSet', createIPSet_format - The format of the file that contains the IPSet.

$sel:location:CreateIPSet', createIPSet_location - The URI of the file that contains the IPSet. For example: https://s3.us-west-2.amazonaws.com/my-bucket/my-object-key.

$sel:activate:CreateIPSet', createIPSet_activate - A Boolean value that indicates whether GuardDuty is to start using the uploaded IPSet.

Request Lenses

createIPSet_clientToken :: Lens' CreateIPSet (Maybe Text)

The idempotency token for the create request.

createIPSet_tags :: Lens' CreateIPSet (Maybe (HashMap Text Text))

The tags to be added to a new IP set resource.

createIPSet_detectorId :: Lens' CreateIPSet Text

The unique ID of the detector of the GuardDuty account that you want to create an IPSet for.

createIPSet_name :: Lens' CreateIPSet Text

The user-friendly name to identify the IPSet.

Allowed characters are alphanumerics, spaces, hyphens (-), and underscores (_).

createIPSet_format :: Lens' CreateIPSet IpSetFormat

The format of the file that contains the IPSet.

createIPSet_location :: Lens' CreateIPSet Text

The URI of the file that contains the IPSet. For example: https://s3.us-west-2.amazonaws.com/my-bucket/my-object-key.

createIPSet_activate :: Lens' CreateIPSet Bool

A Boolean value that indicates whether GuardDuty is to start using the uploaded IPSet.

Destructuring the Response

data CreateIPSetResponse

See: newCreateIPSetResponse smart constructor.

Constructors

CreateIPSetResponse'

Fields

httpStatus :: Int

The response's http status code.

ipSetId :: Text

The ID of the IPSet resource.

newCreateIPSetResponse

Arguments

:: Int
$sel:httpStatus:CreateIPSetResponse'

-> Text
$sel:ipSetId:CreateIPSetResponse'

-> CreateIPSetResponse

Create a value of CreateIPSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateIPSetResponse', createIPSetResponse_httpStatus - The response's http status code.

$sel:ipSetId:CreateIPSetResponse', createIPSetResponse_ipSetId - The ID of the IPSet resource.

Response Lenses

createIPSetResponse_httpStatus :: Lens' CreateIPSetResponse Int

The response's http status code.

createIPSetResponse_ipSetId :: Lens' CreateIPSetResponse Text

The ID of the IPSet resource.