Network.AWS.GuardDuty.DeleteIPSet

Description

Deletes the IPSet specified by the ipSetId. IPSets are called trusted IP lists in the console user interface.

Creating a Request

data DeleteIPSet

See: newDeleteIPSet smart constructor.

Constructors

DeleteIPSet'

Fields

detectorId :: Text

The unique ID of the detector associated with the IPSet.

ipSetId :: Text

The unique ID of the IPSet to delete.

newDeleteIPSet

Arguments

:: Text
$sel:detectorId:DeleteIPSet'

-> Text
$sel:ipSetId:DeleteIPSet'

-> DeleteIPSet

Create a value of DeleteIPSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:detectorId:DeleteIPSet', deleteIPSet_detectorId - The unique ID of the detector associated with the IPSet.

$sel:ipSetId:DeleteIPSet', deleteIPSet_ipSetId - The unique ID of the IPSet to delete.

Request Lenses

deleteIPSet_detectorId :: Lens' DeleteIPSet Text

The unique ID of the detector associated with the IPSet.

deleteIPSet_ipSetId :: Lens' DeleteIPSet Text

The unique ID of the IPSet to delete.

Destructuring the Response

data DeleteIPSetResponse

See: newDeleteIPSetResponse smart constructor.

Constructors

DeleteIPSetResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteIPSetResponse

Arguments

:: Int
$sel:httpStatus:DeleteIPSetResponse'

-> DeleteIPSetResponse

Create a value of DeleteIPSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteIPSetResponse', deleteIPSetResponse_httpStatus - The response's http status code.

Response Lenses

deleteIPSetResponse_httpStatus :: Lens' DeleteIPSetResponse Int

The response's http status code.