Network.AWS.GuardDuty.DeleteInvitations

Description

Deletes invitations sent to the current member account by AWS accounts specified by their account IDs.

Creating a Request

data DeleteInvitations

See: newDeleteInvitations smart constructor.

Constructors

DeleteInvitations'

Fields

accountIds :: NonEmpty Text

A list of account IDs of the AWS accounts that sent invitations to the current member account that you want to delete invitations from.

newDeleteInvitations

Arguments

:: NonEmpty Text
$sel:accountIds:DeleteInvitations'

-> DeleteInvitations

Create a value of DeleteInvitations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountIds:DeleteInvitations', deleteInvitations_accountIds - A list of account IDs of the AWS accounts that sent invitations to the current member account that you want to delete invitations from.

Request Lenses

deleteInvitations_accountIds :: Lens' DeleteInvitations (NonEmpty Text)

A list of account IDs of the AWS accounts that sent invitations to the current member account that you want to delete invitations from.

Destructuring the Response

data DeleteInvitationsResponse

See: newDeleteInvitationsResponse smart constructor.

Constructors

DeleteInvitationsResponse'

Fields

httpStatus :: Int

The response's http status code.

unprocessedAccounts :: [UnprocessedAccount]

A list of objects that contain the unprocessed account and a result string that explains why it was unprocessed.

newDeleteInvitationsResponse

Arguments

:: Int
$sel:httpStatus:DeleteInvitationsResponse'

-> DeleteInvitationsResponse

Create a value of DeleteInvitationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteInvitationsResponse', deleteInvitationsResponse_httpStatus - The response's http status code.

$sel:unprocessedAccounts:DeleteInvitationsResponse', deleteInvitationsResponse_unprocessedAccounts - A list of objects that contain the unprocessed account and a result string that explains why it was unprocessed.

Response Lenses

deleteInvitationsResponse_httpStatus :: Lens' DeleteInvitationsResponse Int

The response's http status code.

deleteInvitationsResponse_unprocessedAccounts :: Lens' DeleteInvitationsResponse [UnprocessedAccount]

A list of objects that contain the unprocessed account and a result string that explains why it was unprocessed.