Network.AWS.GuardDuty.DeleteThreatIntelSet

Description

Deletes the ThreatIntelSet specified by the ThreatIntelSet ID.

Creating a Request

data DeleteThreatIntelSet

See: newDeleteThreatIntelSet smart constructor.

Constructors

DeleteThreatIntelSet'

Fields

detectorId :: Text

The unique ID of the detector that the threatIntelSet is associated with.

threatIntelSetId :: Text

The unique ID of the threatIntelSet that you want to delete.

newDeleteThreatIntelSet

Arguments

:: Text
$sel:detectorId:DeleteThreatIntelSet'

-> Text
$sel:threatIntelSetId:DeleteThreatIntelSet'

-> DeleteThreatIntelSet

Create a value of DeleteThreatIntelSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:detectorId:DeleteThreatIntelSet', deleteThreatIntelSet_detectorId - The unique ID of the detector that the threatIntelSet is associated with.

$sel:threatIntelSetId:DeleteThreatIntelSet', deleteThreatIntelSet_threatIntelSetId - The unique ID of the threatIntelSet that you want to delete.

Request Lenses

deleteThreatIntelSet_detectorId :: Lens' DeleteThreatIntelSet Text

The unique ID of the detector that the threatIntelSet is associated with.

deleteThreatIntelSet_threatIntelSetId :: Lens' DeleteThreatIntelSet Text

The unique ID of the threatIntelSet that you want to delete.

Destructuring the Response

data DeleteThreatIntelSetResponse

See: newDeleteThreatIntelSetResponse smart constructor.

Constructors

DeleteThreatIntelSetResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteThreatIntelSetResponse

Arguments

:: Int
$sel:httpStatus:DeleteThreatIntelSetResponse'

-> DeleteThreatIntelSetResponse

Create a value of DeleteThreatIntelSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteThreatIntelSetResponse', deleteThreatIntelSetResponse_httpStatus - The response's http status code.

Response Lenses

deleteThreatIntelSetResponse_httpStatus :: Lens' DeleteThreatIntelSetResponse Int

The response's http status code.