Network.AWS.GuardDuty.GetDetector

Description

Retrieves an Amazon GuardDuty detector specified by the detectorId.

Creating a Request

data GetDetector

See: newGetDetector smart constructor.

Constructors

GetDetector'

Fields

detectorId :: Text

The unique ID of the detector that you want to get.

newGetDetector

Arguments

:: Text
$sel:detectorId:GetDetector'

-> GetDetector

Create a value of GetDetector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:detectorId:GetDetector', getDetector_detectorId - The unique ID of the detector that you want to get.

Request Lenses

getDetector_detectorId :: Lens' GetDetector Text

The unique ID of the detector that you want to get.

Destructuring the Response

data GetDetectorResponse

See: newGetDetectorResponse smart constructor.

Constructors

GetDetectorResponse'

Fields

dataSources :: Maybe DataSourceConfigurationsResult

Describes which data sources are enabled for the detector.

updatedAt :: Maybe Text

The last-updated timestamp for the detector.

findingPublishingFrequency :: Maybe FindingPublishingFrequency

The publishing frequency of the finding.

createdAt :: Maybe Text

The timestamp of when the detector was created.

tags :: Maybe (HashMap Text Text)

The tags of the detector resource.

httpStatus :: Int

The response's http status code.

serviceRole :: Text

The GuardDuty service role.

status :: DetectorStatus

The detector status.

newGetDetectorResponse

Arguments

:: Int
$sel:httpStatus:GetDetectorResponse'

-> Text
$sel:serviceRole:GetDetectorResponse'

-> DetectorStatus
$sel:status:GetDetectorResponse'

-> GetDetectorResponse

Create a value of GetDetectorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataSources:GetDetectorResponse', getDetectorResponse_dataSources - Describes which data sources are enabled for the detector.

$sel:updatedAt:GetDetectorResponse', getDetectorResponse_updatedAt - The last-updated timestamp for the detector.

$sel:findingPublishingFrequency:GetDetectorResponse', getDetectorResponse_findingPublishingFrequency - The publishing frequency of the finding.

$sel:createdAt:GetDetectorResponse', getDetectorResponse_createdAt - The timestamp of when the detector was created.

$sel:tags:GetDetectorResponse', getDetectorResponse_tags - The tags of the detector resource.

$sel:httpStatus:GetDetectorResponse', getDetectorResponse_httpStatus - The response's http status code.

$sel:serviceRole:GetDetectorResponse', getDetectorResponse_serviceRole - The GuardDuty service role.

$sel:status:GetDetectorResponse', getDetectorResponse_status - The detector status.

Response Lenses

getDetectorResponse_dataSources :: Lens' GetDetectorResponse (Maybe DataSourceConfigurationsResult)

Describes which data sources are enabled for the detector.

getDetectorResponse_updatedAt :: Lens' GetDetectorResponse (Maybe Text)

The last-updated timestamp for the detector.

getDetectorResponse_findingPublishingFrequency :: Lens' GetDetectorResponse (Maybe FindingPublishingFrequency)

The publishing frequency of the finding.

getDetectorResponse_createdAt :: Lens' GetDetectorResponse (Maybe Text)

The timestamp of when the detector was created.

getDetectorResponse_tags :: Lens' GetDetectorResponse (Maybe (HashMap Text Text))

The tags of the detector resource.

getDetectorResponse_httpStatus :: Lens' GetDetectorResponse Int

The response's http status code.

getDetectorResponse_serviceRole :: Lens' GetDetectorResponse Text

The GuardDuty service role.

getDetectorResponse_status :: Lens' GetDetectorResponse DetectorStatus

The detector status.