Network.AWS.GuardDuty.GetFilter

Description

Returns the details of the filter specified by the filter name.

Creating a Request

data GetFilter

See: newGetFilter smart constructor.

Constructors

GetFilter'

Fields

detectorId :: Text

The unique ID of the detector that the filter is associated with.

filterName :: Text

The name of the filter you want to get.

newGetFilter

Arguments

:: Text
$sel:detectorId:GetFilter'

-> Text
$sel:filterName:GetFilter'

-> GetFilter

Create a value of GetFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:detectorId:GetFilter', getFilter_detectorId - The unique ID of the detector that the filter is associated with.

$sel:filterName:GetFilter', getFilter_filterName - The name of the filter you want to get.

Request Lenses

getFilter_detectorId :: Lens' GetFilter Text

The unique ID of the detector that the filter is associated with.

getFilter_filterName :: Lens' GetFilter Text

The name of the filter you want to get.

Destructuring the Response

data GetFilterResponse

See: newGetFilterResponse smart constructor.

Constructors

GetFilterResponse'

Fields

rank :: Maybe Natural

Specifies the position of the filter in the list of current filters. Also specifies the order in which this filter is applied to the findings.

tags :: Maybe (HashMap Text Text)

The tags of the filter resource.

description :: Maybe Text

The description of the filter.

httpStatus :: Int

The response's http status code.

name :: Text

The name of the filter.

action :: FilterAction

Specifies the action that is to be applied to the findings that match the filter.

findingCriteria :: FindingCriteria

Represents the criteria to be used in the filter for querying findings.

newGetFilterResponse

Arguments

:: Int
$sel:httpStatus:GetFilterResponse'

-> Text
$sel:name:GetFilterResponse'

-> FilterAction
$sel:action:GetFilterResponse'

-> FindingCriteria
$sel:findingCriteria:GetFilterResponse'

-> GetFilterResponse

Create a value of GetFilterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rank:GetFilterResponse', getFilterResponse_rank - Specifies the position of the filter in the list of current filters. Also specifies the order in which this filter is applied to the findings.

$sel:tags:GetFilterResponse', getFilterResponse_tags - The tags of the filter resource.

$sel:description:GetFilterResponse', getFilterResponse_description - The description of the filter.

$sel:httpStatus:GetFilterResponse', getFilterResponse_httpStatus - The response's http status code.

$sel:name:GetFilterResponse', getFilterResponse_name - The name of the filter.

$sel:action:GetFilterResponse', getFilterResponse_action - Specifies the action that is to be applied to the findings that match the filter.

$sel:findingCriteria:GetFilterResponse', getFilterResponse_findingCriteria - Represents the criteria to be used in the filter for querying findings.

Response Lenses

getFilterResponse_rank :: Lens' GetFilterResponse (Maybe Natural)

Specifies the position of the filter in the list of current filters. Also specifies the order in which this filter is applied to the findings.

getFilterResponse_tags :: Lens' GetFilterResponse (Maybe (HashMap Text Text))

The tags of the filter resource.

getFilterResponse_description :: Lens' GetFilterResponse (Maybe Text)

The description of the filter.

getFilterResponse_httpStatus :: Lens' GetFilterResponse Int

The response's http status code.

getFilterResponse_name :: Lens' GetFilterResponse Text

The name of the filter.

getFilterResponse_action :: Lens' GetFilterResponse FilterAction

Specifies the action that is to be applied to the findings that match the filter.

getFilterResponse_findingCriteria :: Lens' GetFilterResponse FindingCriteria

Represents the criteria to be used in the filter for querying findings.