Network.AWS.GuardDuty.GetMasterAccount

Description

Provides the details for the GuardDuty administrator account associated with the current GuardDuty member account.

Creating a Request

data GetMasterAccount

See: newGetMasterAccount smart constructor.

Constructors

GetMasterAccount'

Fields

detectorId :: Text

The unique ID of the detector of the GuardDuty member account.

newGetMasterAccount

Arguments

:: Text
$sel:detectorId:GetMasterAccount'

-> GetMasterAccount

Create a value of GetMasterAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:detectorId:GetMasterAccount', getMasterAccount_detectorId - The unique ID of the detector of the GuardDuty member account.

Request Lenses

getMasterAccount_detectorId :: Lens' GetMasterAccount Text

The unique ID of the detector of the GuardDuty member account.

Destructuring the Response

data GetMasterAccountResponse

See: newGetMasterAccountResponse smart constructor.

Constructors

GetMasterAccountResponse'

Fields

httpStatus :: Int

The response's http status code.

master :: Master

The administrator account details.

newGetMasterAccountResponse

Arguments

:: Int
$sel:httpStatus:GetMasterAccountResponse'

-> Master
$sel:master:GetMasterAccountResponse'

-> GetMasterAccountResponse

Create a value of GetMasterAccountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetMasterAccountResponse', getMasterAccountResponse_httpStatus - The response's http status code.

$sel:master:GetMasterAccountResponse', getMasterAccountResponse_master - The administrator account details.

Response Lenses

getMasterAccountResponse_httpStatus :: Lens' GetMasterAccountResponse Int

The response's http status code.

getMasterAccountResponse_master :: Lens' GetMasterAccountResponse Master

The administrator account details.