Network.AWS.GuardDuty.GetUsageStatistics

Description

Lists Amazon GuardDuty usage statistics over the last 30 days for the specified detector ID. For newly enabled detectors or data sources the cost returned will include only the usage so far under 30 days, this may differ from the cost metrics in the console, which projects usage over 30 days to provide a monthly cost estimate. For more information see Understanding How Usage Costs are Calculated.

Creating a Request

data GetUsageStatistics

See: newGetUsageStatistics smart constructor.

Constructors

GetUsageStatistics'

Fields

nextToken :: Maybe Text

A token to use for paginating results that are returned in the response. Set the value of this parameter to null for the first request to a list action. For subsequent calls, use the NextToken value returned from the previous request to continue listing results after the first page.

unit :: Maybe Text

The currency unit you would like to view your usage statistics in. Current valid values are USD.

maxResults :: Maybe Natural

The maximum number of results to return in the response.

detectorId :: Text

The ID of the detector that specifies the GuardDuty service whose usage statistics you want to retrieve.

usageStatisticType :: UsageStatisticType

The type of usage statistics to retrieve.

usageCriteria :: UsageCriteria

Represents the criteria used for querying usage.

newGetUsageStatistics

Arguments

:: Text
$sel:detectorId:GetUsageStatistics'

-> UsageStatisticType
$sel:usageStatisticType:GetUsageStatistics'

-> UsageCriteria
$sel:usageCriteria:GetUsageStatistics'

-> GetUsageStatistics

Create a value of GetUsageStatistics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetUsageStatistics', getUsageStatistics_nextToken - A token to use for paginating results that are returned in the response. Set the value of this parameter to null for the first request to a list action. For subsequent calls, use the NextToken value returned from the previous request to continue listing results after the first page.

$sel:unit:GetUsageStatistics', getUsageStatistics_unit - The currency unit you would like to view your usage statistics in. Current valid values are USD.

$sel:maxResults:GetUsageStatistics', getUsageStatistics_maxResults - The maximum number of results to return in the response.

$sel:detectorId:GetUsageStatistics', getUsageStatistics_detectorId - The ID of the detector that specifies the GuardDuty service whose usage statistics you want to retrieve.

$sel:usageStatisticType:GetUsageStatistics', getUsageStatistics_usageStatisticType - The type of usage statistics to retrieve.

$sel:usageCriteria:GetUsageStatistics', getUsageStatistics_usageCriteria - Represents the criteria used for querying usage.

Request Lenses

getUsageStatistics_nextToken :: Lens' GetUsageStatistics (Maybe Text)

A token to use for paginating results that are returned in the response. Set the value of this parameter to null for the first request to a list action. For subsequent calls, use the NextToken value returned from the previous request to continue listing results after the first page.

getUsageStatistics_unit :: Lens' GetUsageStatistics (Maybe Text)

The currency unit you would like to view your usage statistics in. Current valid values are USD.

getUsageStatistics_maxResults :: Lens' GetUsageStatistics (Maybe Natural)

The maximum number of results to return in the response.

getUsageStatistics_detectorId :: Lens' GetUsageStatistics Text

The ID of the detector that specifies the GuardDuty service whose usage statistics you want to retrieve.

getUsageStatistics_usageStatisticType :: Lens' GetUsageStatistics UsageStatisticType

The type of usage statistics to retrieve.

getUsageStatistics_usageCriteria :: Lens' GetUsageStatistics UsageCriteria

Represents the criteria used for querying usage.

Destructuring the Response

data GetUsageStatisticsResponse

See: newGetUsageStatisticsResponse smart constructor.

Constructors

GetUsageStatisticsResponse'

Fields

usageStatistics :: Maybe UsageStatistics

The usage statistics object. If a UsageStatisticType was provided, the objects representing other types will be null.

nextToken :: Maybe Text

The pagination parameter to be used on the next list operation to retrieve more items.

httpStatus :: Int

The response's http status code.

newGetUsageStatisticsResponse

Arguments

:: Int
$sel:httpStatus:GetUsageStatisticsResponse'

-> GetUsageStatisticsResponse

Create a value of GetUsageStatisticsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:usageStatistics:GetUsageStatisticsResponse', getUsageStatisticsResponse_usageStatistics - The usage statistics object. If a UsageStatisticType was provided, the objects representing other types will be null.

$sel:nextToken:GetUsageStatistics', getUsageStatisticsResponse_nextToken - The pagination parameter to be used on the next list operation to retrieve more items.

$sel:httpStatus:GetUsageStatisticsResponse', getUsageStatisticsResponse_httpStatus - The response's http status code.

Response Lenses

getUsageStatisticsResponse_usageStatistics :: Lens' GetUsageStatisticsResponse (Maybe UsageStatistics)

The usage statistics object. If a UsageStatisticType was provided, the objects representing other types will be null.

getUsageStatisticsResponse_nextToken :: Lens' GetUsageStatisticsResponse (Maybe Text)

The pagination parameter to be used on the next list operation to retrieve more items.

getUsageStatisticsResponse_httpStatus :: Lens' GetUsageStatisticsResponse Int

The response's http status code.