Network.AWS.GuardDuty.ListFindings

Description

Lists Amazon GuardDuty findings for the specified detector ID.

This operation returns paginated results.

Creating a Request

data ListFindings

See: newListFindings smart constructor.

Constructors

ListFindings'

Fields

nextToken :: Maybe Text

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the list action. For subsequent calls to the action, fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

sortCriteria :: Maybe SortCriteria

Represents the criteria used for sorting findings.

maxResults :: Maybe Natural

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 50. The maximum value is 50.

findingCriteria :: Maybe FindingCriteria

Represents the criteria used for querying findings. Valid values include:

JSON field name

accountId

region

confidence

id

resource.accessKeyDetails.accessKeyId

resource.accessKeyDetails.principalId

resource.accessKeyDetails.userName

resource.accessKeyDetails.userType

resource.instanceDetails.iamInstanceProfile.id

resource.instanceDetails.imageId

resource.instanceDetails.instanceId

resource.instanceDetails.networkInterfaces.ipv6Addresses

resource.instanceDetails.networkInterfaces.privateIpAddresses.privateIpAddress

resource.instanceDetails.networkInterfaces.publicDnsName

resource.instanceDetails.networkInterfaces.publicIp

resource.instanceDetails.networkInterfaces.securityGroups.groupId

resource.instanceDetails.networkInterfaces.securityGroups.groupName

resource.instanceDetails.networkInterfaces.subnetId

resource.instanceDetails.networkInterfaces.vpcId

resource.instanceDetails.tags.key

resource.instanceDetails.tags.value

resource.resourceType

service.action.actionType

service.action.awsApiCallAction.api

service.action.awsApiCallAction.callerType

service.action.awsApiCallAction.remoteIpDetails.city.cityName

service.action.awsApiCallAction.remoteIpDetails.country.countryName

service.action.awsApiCallAction.remoteIpDetails.ipAddressV4

service.action.awsApiCallAction.remoteIpDetails.organization.asn

service.action.awsApiCallAction.remoteIpDetails.organization.asnOrg

service.action.awsApiCallAction.serviceName

service.action.dnsRequestAction.domain

service.action.networkConnectionAction.blocked

service.action.networkConnectionAction.connectionDirection

service.action.networkConnectionAction.localPortDetails.port

service.action.networkConnectionAction.protocol

service.action.networkConnectionAction.remoteIpDetails.city.cityName

service.action.networkConnectionAction.remoteIpDetails.country.countryName

service.action.networkConnectionAction.remoteIpDetails.ipAddressV4

service.action.networkConnectionAction.remoteIpDetails.organization.asn

service.action.networkConnectionAction.remoteIpDetails.organization.asnOrg

service.action.networkConnectionAction.remotePortDetails.port

service.additionalInfo.threatListName

service.archived

When this attribute is set to 'true', only archived findings are listed. When it's set to 'false', only unarchived findings are listed. When this attribute is not set, all existing findings are listed.

service.resourceRole

severity

type

updatedAt

Type: Timestamp in Unix Epoch millisecond format: 1486685375000

detectorId :: Text

The ID of the detector that specifies the GuardDuty service whose findings you want to list.

newListFindings

Arguments

:: Text
$sel:detectorId:ListFindings'

-> ListFindings

Create a value of ListFindings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFindings', listFindings_nextToken - You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the list action. For subsequent calls to the action, fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

$sel:sortCriteria:ListFindings', listFindings_sortCriteria - Represents the criteria used for sorting findings.

$sel:maxResults:ListFindings', listFindings_maxResults - You can use this parameter to indicate the maximum number of items you want in the response. The default value is 50. The maximum value is 50.

$sel:findingCriteria:ListFindings', listFindings_findingCriteria - Represents the criteria used for querying findings. Valid values include:

JSON field name

accountId

region

confidence

id

resource.accessKeyDetails.accessKeyId

resource.accessKeyDetails.principalId

resource.accessKeyDetails.userName

resource.accessKeyDetails.userType

resource.instanceDetails.iamInstanceProfile.id

resource.instanceDetails.imageId

resource.instanceDetails.instanceId

resource.instanceDetails.networkInterfaces.ipv6Addresses

resource.instanceDetails.networkInterfaces.privateIpAddresses.privateIpAddress

resource.instanceDetails.networkInterfaces.publicDnsName

resource.instanceDetails.networkInterfaces.publicIp

resource.instanceDetails.networkInterfaces.securityGroups.groupId

resource.instanceDetails.networkInterfaces.securityGroups.groupName

resource.instanceDetails.networkInterfaces.subnetId

resource.instanceDetails.networkInterfaces.vpcId

resource.instanceDetails.tags.key

resource.instanceDetails.tags.value

resource.resourceType

service.action.actionType

service.action.awsApiCallAction.api

service.action.awsApiCallAction.callerType

service.action.awsApiCallAction.remoteIpDetails.city.cityName

service.action.awsApiCallAction.remoteIpDetails.country.countryName

service.action.awsApiCallAction.remoteIpDetails.ipAddressV4

service.action.awsApiCallAction.remoteIpDetails.organization.asn

service.action.awsApiCallAction.remoteIpDetails.organization.asnOrg

service.action.awsApiCallAction.serviceName

service.action.dnsRequestAction.domain

service.action.networkConnectionAction.blocked

service.action.networkConnectionAction.connectionDirection

service.action.networkConnectionAction.localPortDetails.port

service.action.networkConnectionAction.protocol

service.action.networkConnectionAction.remoteIpDetails.city.cityName

service.action.networkConnectionAction.remoteIpDetails.country.countryName

service.action.networkConnectionAction.remoteIpDetails.ipAddressV4

service.action.networkConnectionAction.remoteIpDetails.organization.asn

service.action.networkConnectionAction.remoteIpDetails.organization.asnOrg

service.action.networkConnectionAction.remotePortDetails.port

service.additionalInfo.threatListName

service.archived

When this attribute is set to 'true', only archived findings are listed. When it's set to 'false', only unarchived findings are listed. When this attribute is not set, all existing findings are listed.

service.resourceRole

severity

type

updatedAt

Type: Timestamp in Unix Epoch millisecond format: 1486685375000

$sel:detectorId:ListFindings', listFindings_detectorId - The ID of the detector that specifies the GuardDuty service whose findings you want to list.

Request Lenses

listFindings_nextToken :: Lens' ListFindings (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the list action. For subsequent calls to the action, fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

listFindings_sortCriteria :: Lens' ListFindings (Maybe SortCriteria)

Represents the criteria used for sorting findings.

listFindings_maxResults :: Lens' ListFindings (Maybe Natural)

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 50. The maximum value is 50.

listFindings_findingCriteria :: Lens' ListFindings (Maybe FindingCriteria)

Represents the criteria used for querying findings. Valid values include:

JSON field name

accountId

region

confidence

id

resource.accessKeyDetails.accessKeyId

resource.accessKeyDetails.principalId

resource.accessKeyDetails.userName

resource.accessKeyDetails.userType

resource.instanceDetails.iamInstanceProfile.id

resource.instanceDetails.imageId

resource.instanceDetails.instanceId

resource.instanceDetails.networkInterfaces.ipv6Addresses

resource.instanceDetails.networkInterfaces.privateIpAddresses.privateIpAddress

resource.instanceDetails.networkInterfaces.publicDnsName

resource.instanceDetails.networkInterfaces.publicIp

resource.instanceDetails.networkInterfaces.securityGroups.groupId

resource.instanceDetails.networkInterfaces.securityGroups.groupName

resource.instanceDetails.networkInterfaces.subnetId

resource.instanceDetails.networkInterfaces.vpcId

resource.instanceDetails.tags.key

resource.instanceDetails.tags.value

resource.resourceType

service.action.actionType

service.action.awsApiCallAction.api

service.action.awsApiCallAction.callerType

service.action.awsApiCallAction.remoteIpDetails.city.cityName

service.action.awsApiCallAction.remoteIpDetails.country.countryName

service.action.awsApiCallAction.remoteIpDetails.ipAddressV4

service.action.awsApiCallAction.remoteIpDetails.organization.asn

service.action.awsApiCallAction.remoteIpDetails.organization.asnOrg

service.action.awsApiCallAction.serviceName

service.action.dnsRequestAction.domain

service.action.networkConnectionAction.blocked

service.action.networkConnectionAction.connectionDirection

service.action.networkConnectionAction.localPortDetails.port

service.action.networkConnectionAction.protocol

service.action.networkConnectionAction.remoteIpDetails.city.cityName

service.action.networkConnectionAction.remoteIpDetails.country.countryName

service.action.networkConnectionAction.remoteIpDetails.ipAddressV4

service.action.networkConnectionAction.remoteIpDetails.organization.asn

service.action.networkConnectionAction.remoteIpDetails.organization.asnOrg

service.action.networkConnectionAction.remotePortDetails.port

service.additionalInfo.threatListName

service.archived

When this attribute is set to 'true', only archived findings are listed. When it's set to 'false', only unarchived findings are listed. When this attribute is not set, all existing findings are listed.

service.resourceRole

severity

type

updatedAt

Type: Timestamp in Unix Epoch millisecond format: 1486685375000

listFindings_detectorId :: Lens' ListFindings Text

The ID of the detector that specifies the GuardDuty service whose findings you want to list.

Destructuring the Response

data ListFindingsResponse

See: newListFindingsResponse smart constructor.

Constructors

ListFindingsResponse'

Fields

nextToken :: Maybe Text

The pagination parameter to be used on the next list operation to retrieve more items.

httpStatus :: Int

The response's http status code.

findingIds :: [Text]

The IDs of the findings that you're listing.

newListFindingsResponse

Arguments

:: Int
$sel:httpStatus:ListFindingsResponse'

-> ListFindingsResponse

Create a value of ListFindingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFindings', listFindingsResponse_nextToken - The pagination parameter to be used on the next list operation to retrieve more items.

$sel:httpStatus:ListFindingsResponse', listFindingsResponse_httpStatus - The response's http status code.

$sel:findingIds:ListFindingsResponse', listFindingsResponse_findingIds - The IDs of the findings that you're listing.

Response Lenses

listFindingsResponse_nextToken :: Lens' ListFindingsResponse (Maybe Text)

The pagination parameter to be used on the next list operation to retrieve more items.

listFindingsResponse_httpStatus :: Lens' ListFindingsResponse Int

The response's http status code.

listFindingsResponse_findingIds :: Lens' ListFindingsResponse [Text]

The IDs of the findings that you're listing.