Network.AWS.GuardDuty.ListIPSets

Description

Lists the IPSets of the GuardDuty service specified by the detector ID. If you use this operation from a member account, the IPSets returned are the IPSets from the associated administrator account.

This operation returns paginated results.

Creating a Request

data ListIPSets

See: newListIPSets smart constructor.

Constructors

ListIPSets'

Fields

nextToken :: Maybe Text

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the list action. For subsequent calls to the action, fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

maxResults :: Maybe Natural

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 50. The maximum value is 50.

detectorId :: Text

The unique ID of the detector that the IPSet is associated with.

newListIPSets

Arguments

:: Text
$sel:detectorId:ListIPSets'

-> ListIPSets

Create a value of ListIPSets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListIPSets', listIPSets_nextToken - You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the list action. For subsequent calls to the action, fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

$sel:maxResults:ListIPSets', listIPSets_maxResults - You can use this parameter to indicate the maximum number of items you want in the response. The default value is 50. The maximum value is 50.

$sel:detectorId:ListIPSets', listIPSets_detectorId - The unique ID of the detector that the IPSet is associated with.

Request Lenses

listIPSets_nextToken :: Lens' ListIPSets (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the list action. For subsequent calls to the action, fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

listIPSets_maxResults :: Lens' ListIPSets (Maybe Natural)

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 50. The maximum value is 50.

listIPSets_detectorId :: Lens' ListIPSets Text

The unique ID of the detector that the IPSet is associated with.

Destructuring the Response

data ListIPSetsResponse

See: newListIPSetsResponse smart constructor.

Constructors

ListIPSetsResponse'

Fields

nextToken :: Maybe Text

The pagination parameter to be used on the next list operation to retrieve more items.

httpStatus :: Int

The response's http status code.

ipSetIds :: [Text]

The IDs of the IPSet resources.

newListIPSetsResponse

Arguments

:: Int
$sel:httpStatus:ListIPSetsResponse'

-> ListIPSetsResponse

Create a value of ListIPSetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListIPSets', listIPSetsResponse_nextToken - The pagination parameter to be used on the next list operation to retrieve more items.

$sel:httpStatus:ListIPSetsResponse', listIPSetsResponse_httpStatus - The response's http status code.

$sel:ipSetIds:ListIPSetsResponse', listIPSetsResponse_ipSetIds - The IDs of the IPSet resources.

Response Lenses

listIPSetsResponse_nextToken :: Lens' ListIPSetsResponse (Maybe Text)

The pagination parameter to be used on the next list operation to retrieve more items.

listIPSetsResponse_httpStatus :: Lens' ListIPSetsResponse Int

The response's http status code.

listIPSetsResponse_ipSetIds :: Lens' ListIPSetsResponse [Text]

The IDs of the IPSet resources.