Network.AWS.GuardDuty.ListInvitations

Description

Lists all GuardDuty membership invitations that were sent to the current AWS account.

This operation returns paginated results.

Creating a Request

data ListInvitations

See: newListInvitations smart constructor.

Constructors

ListInvitations'

Fields

nextToken :: Maybe Text

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the list action. For subsequent calls to the action, fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

maxResults :: Maybe Natural

You can use this parameter to indicate the maximum number of items that you want in the response. The default value is 50. The maximum value is 50.

newListInvitations :: ListInvitations

Create a value of ListInvitations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListInvitations', listInvitations_nextToken - You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the list action. For subsequent calls to the action, fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

$sel:maxResults:ListInvitations', listInvitations_maxResults - You can use this parameter to indicate the maximum number of items that you want in the response. The default value is 50. The maximum value is 50.

Request Lenses

listInvitations_nextToken :: Lens' ListInvitations (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the list action. For subsequent calls to the action, fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

listInvitations_maxResults :: Lens' ListInvitations (Maybe Natural)

You can use this parameter to indicate the maximum number of items that you want in the response. The default value is 50. The maximum value is 50.

Destructuring the Response

data ListInvitationsResponse

See: newListInvitationsResponse smart constructor.

Constructors

ListInvitationsResponse'

Fields

nextToken :: Maybe Text

The pagination parameter to be used on the next list operation to retrieve more items.

invitations :: Maybe [Invitation]

A list of invitation descriptions.

httpStatus :: Int

The response's http status code.

newListInvitationsResponse

Arguments

:: Int
$sel:httpStatus:ListInvitationsResponse'

-> ListInvitationsResponse

Create a value of ListInvitationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListInvitations', listInvitationsResponse_nextToken - The pagination parameter to be used on the next list operation to retrieve more items.

$sel:invitations:ListInvitationsResponse', listInvitationsResponse_invitations - A list of invitation descriptions.

$sel:httpStatus:ListInvitationsResponse', listInvitationsResponse_httpStatus - The response's http status code.

Response Lenses

listInvitationsResponse_nextToken :: Lens' ListInvitationsResponse (Maybe Text)

The pagination parameter to be used on the next list operation to retrieve more items.

listInvitationsResponse_invitations :: Lens' ListInvitationsResponse (Maybe [Invitation])

A list of invitation descriptions.

listInvitationsResponse_httpStatus :: Lens' ListInvitationsResponse Int

The response's http status code.