Network.AWS.GuardDuty.ListMembers

Description

Lists details about all member accounts for the current GuardDuty administrator account.

This operation returns paginated results.

Creating a Request

data ListMembers

See: newListMembers smart constructor.

Constructors

ListMembers'

Fields

onlyAssociated :: Maybe Text

Specifies whether to only return associated members or to return all members (including members who haven't been invited yet or have been disassociated).

nextToken :: Maybe Text

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the list action. For subsequent calls to the action, fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

maxResults :: Maybe Natural

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 50. The maximum value is 50.

detectorId :: Text

The unique ID of the detector the member is associated with.

newListMembers

Arguments

:: Text
$sel:detectorId:ListMembers'

-> ListMembers

Create a value of ListMembers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:onlyAssociated:ListMembers', listMembers_onlyAssociated - Specifies whether to only return associated members or to return all members (including members who haven't been invited yet or have been disassociated).

$sel:nextToken:ListMembers', listMembers_nextToken - You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the list action. For subsequent calls to the action, fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

$sel:maxResults:ListMembers', listMembers_maxResults - You can use this parameter to indicate the maximum number of items you want in the response. The default value is 50. The maximum value is 50.

$sel:detectorId:ListMembers', listMembers_detectorId - The unique ID of the detector the member is associated with.

Request Lenses

listMembers_onlyAssociated :: Lens' ListMembers (Maybe Text)

Specifies whether to only return associated members or to return all members (including members who haven't been invited yet or have been disassociated).

listMembers_nextToken :: Lens' ListMembers (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the list action. For subsequent calls to the action, fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

listMembers_maxResults :: Lens' ListMembers (Maybe Natural)

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 50. The maximum value is 50.

listMembers_detectorId :: Lens' ListMembers Text

The unique ID of the detector the member is associated with.

Destructuring the Response

data ListMembersResponse

See: newListMembersResponse smart constructor.

Constructors

ListMembersResponse'

Fields

members :: Maybe [Member]

A list of members.

nextToken :: Maybe Text

The pagination parameter to be used on the next list operation to retrieve more items.

httpStatus :: Int

The response's http status code.

newListMembersResponse

Arguments

:: Int
$sel:httpStatus:ListMembersResponse'

-> ListMembersResponse

Create a value of ListMembersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:members:ListMembersResponse', listMembersResponse_members - A list of members.

$sel:nextToken:ListMembers', listMembersResponse_nextToken - The pagination parameter to be used on the next list operation to retrieve more items.

$sel:httpStatus:ListMembersResponse', listMembersResponse_httpStatus - The response's http status code.

Response Lenses

listMembersResponse_members :: Lens' ListMembersResponse (Maybe [Member])

A list of members.

listMembersResponse_nextToken :: Lens' ListMembersResponse (Maybe Text)

The pagination parameter to be used on the next list operation to retrieve more items.

listMembersResponse_httpStatus :: Lens' ListMembersResponse Int

The response's http status code.