Network.AWS.GuardDuty.StartMonitoringMembers

Description

Turns on GuardDuty monitoring of the specified member accounts. Use this operation to restart monitoring of accounts that you stopped monitoring with the StopMonitoringMembers operation.

Creating a Request

data StartMonitoringMembers

See: newStartMonitoringMembers smart constructor.

Constructors

StartMonitoringMembers'

Fields

detectorId :: Text

The unique ID of the detector of the GuardDuty administrator account associated with the member accounts to monitor.

accountIds :: NonEmpty Text

A list of account IDs of the GuardDuty member accounts to start monitoring.

newStartMonitoringMembers

Arguments

:: Text
$sel:detectorId:StartMonitoringMembers'

-> NonEmpty Text
$sel:accountIds:StartMonitoringMembers'

-> StartMonitoringMembers

Create a value of StartMonitoringMembers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:detectorId:StartMonitoringMembers', startMonitoringMembers_detectorId - The unique ID of the detector of the GuardDuty administrator account associated with the member accounts to monitor.

$sel:accountIds:StartMonitoringMembers', startMonitoringMembers_accountIds - A list of account IDs of the GuardDuty member accounts to start monitoring.

Request Lenses

startMonitoringMembers_detectorId :: Lens' StartMonitoringMembers Text

The unique ID of the detector of the GuardDuty administrator account associated with the member accounts to monitor.

startMonitoringMembers_accountIds :: Lens' StartMonitoringMembers (NonEmpty Text)

A list of account IDs of the GuardDuty member accounts to start monitoring.

Destructuring the Response

data StartMonitoringMembersResponse

See: newStartMonitoringMembersResponse smart constructor.

Constructors

StartMonitoringMembersResponse'

Fields

httpStatus :: Int

The response's http status code.

unprocessedAccounts :: [UnprocessedAccount]

A list of objects that contain the unprocessed account and a result string that explains why it was unprocessed.

newStartMonitoringMembersResponse

Arguments

:: Int
$sel:httpStatus:StartMonitoringMembersResponse'

-> StartMonitoringMembersResponse

Create a value of StartMonitoringMembersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StartMonitoringMembersResponse', startMonitoringMembersResponse_httpStatus - The response's http status code.

$sel:unprocessedAccounts:StartMonitoringMembersResponse', startMonitoringMembersResponse_unprocessedAccounts - A list of objects that contain the unprocessed account and a result string that explains why it was unprocessed.

Response Lenses

startMonitoringMembersResponse_httpStatus :: Lens' StartMonitoringMembersResponse Int

The response's http status code.

startMonitoringMembersResponse_unprocessedAccounts :: Lens' StartMonitoringMembersResponse [UnprocessedAccount]

A list of objects that contain the unprocessed account and a result string that explains why it was unprocessed.