Network.AWS.GuardDuty.Types.BucketLevelPermissions

Description

Documentation

data BucketLevelPermissions

Contains information about the bucket level permissions for the S3 bucket.

See: newBucketLevelPermissions smart constructor.

Constructors

BucketLevelPermissions'

Fields

blockPublicAccess :: Maybe BlockPublicAccess

Contains information on which account level S3 Block Public Access settings are applied to the S3 bucket.

bucketPolicy :: Maybe BucketPolicy

Contains information on the bucket policies for the S3 bucket.

accessControlList :: Maybe AccessControlList

Contains information on how Access Control Policies are applied to the bucket.

newBucketLevelPermissions :: BucketLevelPermissions

Create a value of BucketLevelPermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blockPublicAccess:BucketLevelPermissions', bucketLevelPermissions_blockPublicAccess - Contains information on which account level S3 Block Public Access settings are applied to the S3 bucket.

$sel:bucketPolicy:BucketLevelPermissions', bucketLevelPermissions_bucketPolicy - Contains information on the bucket policies for the S3 bucket.

$sel:accessControlList:BucketLevelPermissions', bucketLevelPermissions_accessControlList - Contains information on how Access Control Policies are applied to the bucket.

bucketLevelPermissions_blockPublicAccess :: Lens' BucketLevelPermissions (Maybe BlockPublicAccess)

Contains information on which account level S3 Block Public Access settings are applied to the S3 bucket.

bucketLevelPermissions_bucketPolicy :: Lens' BucketLevelPermissions (Maybe BucketPolicy)

Contains information on the bucket policies for the S3 bucket.

bucketLevelPermissions_accessControlList :: Lens' BucketLevelPermissions (Maybe AccessControlList)

Contains information on how Access Control Policies are applied to the bucket.