Network.AWS.GuardDuty.Types.Condition

Description

Documentation

data Condition

Contains information about the condition.

See: newCondition smart constructor.

Constructors

Condition'

Fields

eq :: Maybe [Text]

Represents the equal condition to be applied to a single field when querying for findings.

lessThan :: Maybe Integer

Represents a less than condition to be applied to a single field when querying for findings.

lte :: Maybe Int

Represents a less than or equal condition to be applied to a single field when querying for findings.

greaterThanOrEqual :: Maybe Integer

Represents a greater than or equal condition to be applied to a single field when querying for findings.

lessThanOrEqual :: Maybe Integer

Represents a less than or equal condition to be applied to a single field when querying for findings.

gt :: Maybe Int

Represents a greater than condition to be applied to a single field when querying for findings.

equals :: Maybe [Text]

Represents an equal ____ condition to be applied to a single field when querying for findings.

neq :: Maybe [Text]

Represents the not equal condition to be applied to a single field when querying for findings.

notEquals :: Maybe [Text]

Represents a not equal ____ condition to be applied to a single field when querying for findings.

lt :: Maybe Int

Represents a less than condition to be applied to a single field when querying for findings.

gte :: Maybe Int

Represents a greater than or equal condition to be applied to a single field when querying for findings.

greaterThan :: Maybe Integer

Represents a greater than condition to be applied to a single field when querying for findings.

newCondition :: Condition

Create a value of Condition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eq:Condition', condition_eq - Represents the equal condition to be applied to a single field when querying for findings.

$sel:lessThan:Condition', condition_lessThan - Represents a less than condition to be applied to a single field when querying for findings.

$sel:lte:Condition', condition_lte - Represents a less than or equal condition to be applied to a single field when querying for findings.

$sel:greaterThanOrEqual:Condition', condition_greaterThanOrEqual - Represents a greater than or equal condition to be applied to a single field when querying for findings.

$sel:lessThanOrEqual:Condition', condition_lessThanOrEqual - Represents a less than or equal condition to be applied to a single field when querying for findings.

$sel:gt:Condition', condition_gt - Represents a greater than condition to be applied to a single field when querying for findings.

$sel:equals:Condition', condition_equals - Represents an equal ____ condition to be applied to a single field when querying for findings.

$sel:neq:Condition', condition_neq - Represents the not equal condition to be applied to a single field when querying for findings.

$sel:notEquals:Condition', condition_notEquals - Represents a not equal ____ condition to be applied to a single field when querying for findings.

$sel:lt:Condition', condition_lt - Represents a less than condition to be applied to a single field when querying for findings.

$sel:gte:Condition', condition_gte - Represents a greater than or equal condition to be applied to a single field when querying for findings.

$sel:greaterThan:Condition', condition_greaterThan - Represents a greater than condition to be applied to a single field when querying for findings.

condition_eq :: Lens' Condition (Maybe [Text])

Represents the equal condition to be applied to a single field when querying for findings.

condition_lessThan :: Lens' Condition (Maybe Integer)

Represents a less than condition to be applied to a single field when querying for findings.

condition_lte :: Lens' Condition (Maybe Int)

Represents a less than or equal condition to be applied to a single field when querying for findings.

condition_greaterThanOrEqual :: Lens' Condition (Maybe Integer)

Represents a greater than or equal condition to be applied to a single field when querying for findings.

condition_lessThanOrEqual :: Lens' Condition (Maybe Integer)

Represents a less than or equal condition to be applied to a single field when querying for findings.

condition_gt :: Lens' Condition (Maybe Int)

Represents a greater than condition to be applied to a single field when querying for findings.

condition_equals :: Lens' Condition (Maybe [Text])

Represents an equal ____ condition to be applied to a single field when querying for findings.

condition_neq :: Lens' Condition (Maybe [Text])

Represents the not equal condition to be applied to a single field when querying for findings.

condition_notEquals :: Lens' Condition (Maybe [Text])

Represents a not equal ____ condition to be applied to a single field when querying for findings.

condition_lt :: Lens' Condition (Maybe Int)

Represents a less than condition to be applied to a single field when querying for findings.

condition_gte :: Lens' Condition (Maybe Int)

Represents a greater than or equal condition to be applied to a single field when querying for findings.

condition_greaterThan :: Lens' Condition (Maybe Integer)

Represents a greater than condition to be applied to a single field when querying for findings.