Network.AWS.GuardDuty.Types.DataSourceConfigurations

Description

Documentation

data DataSourceConfigurations

Contains information about which data sources are enabled.

See: newDataSourceConfigurations smart constructor.

Constructors

DataSourceConfigurations'

Fields

s3Logs :: Maybe S3LogsConfiguration

Describes whether S3 data event logs are enabled as a data source.

newDataSourceConfigurations :: DataSourceConfigurations

Create a value of DataSourceConfigurations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Logs:DataSourceConfigurations', dataSourceConfigurations_s3Logs - Describes whether S3 data event logs are enabled as a data source.

dataSourceConfigurations_s3Logs :: Lens' DataSourceConfigurations (Maybe S3LogsConfiguration)

Describes whether S3 data event logs are enabled as a data source.