Network.AWS.GuardDuty.Types.DataSourceConfigurationsResult

Description

Documentation

data DataSourceConfigurationsResult

Contains information on the status of data sources for the detector.

See: newDataSourceConfigurationsResult smart constructor.

Constructors

DataSourceConfigurationsResult'

Fields

cloudTrail :: CloudTrailConfigurationResult

An object that contains information on the status of CloudTrail as a data source.

dNSLogs :: DNSLogsConfigurationResult

An object that contains information on the status of DNS logs as a data source.

flowLogs :: FlowLogsConfigurationResult

An object that contains information on the status of VPC flow logs as a data source.

s3Logs :: S3LogsConfigurationResult

An object that contains information on the status of S3 Data event logs as a data source.

newDataSourceConfigurationsResult

Arguments

:: CloudTrailConfigurationResult
$sel:cloudTrail:DataSourceConfigurationsResult'

-> DNSLogsConfigurationResult
$sel:dNSLogs:DataSourceConfigurationsResult'

-> FlowLogsConfigurationResult
$sel:flowLogs:DataSourceConfigurationsResult'

-> S3LogsConfigurationResult
$sel:s3Logs:DataSourceConfigurationsResult'

-> DataSourceConfigurationsResult

Create a value of DataSourceConfigurationsResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cloudTrail:DataSourceConfigurationsResult', dataSourceConfigurationsResult_cloudTrail - An object that contains information on the status of CloudTrail as a data source.

$sel:dNSLogs:DataSourceConfigurationsResult', dataSourceConfigurationsResult_dNSLogs - An object that contains information on the status of DNS logs as a data source.

$sel:flowLogs:DataSourceConfigurationsResult', dataSourceConfigurationsResult_flowLogs - An object that contains information on the status of VPC flow logs as a data source.

$sel:s3Logs:DataSourceConfigurationsResult', dataSourceConfigurationsResult_s3Logs - An object that contains information on the status of S3 Data event logs as a data source.

dataSourceConfigurationsResult_cloudTrail :: Lens' DataSourceConfigurationsResult CloudTrailConfigurationResult

An object that contains information on the status of CloudTrail as a data source.

dataSourceConfigurationsResult_dNSLogs :: Lens' DataSourceConfigurationsResult DNSLogsConfigurationResult

An object that contains information on the status of DNS logs as a data source.

dataSourceConfigurationsResult_flowLogs :: Lens' DataSourceConfigurationsResult FlowLogsConfigurationResult

An object that contains information on the status of VPC flow logs as a data source.

dataSourceConfigurationsResult_s3Logs :: Lens' DataSourceConfigurationsResult S3LogsConfigurationResult

An object that contains information on the status of S3 Data event logs as a data source.