Network.AWS.GuardDuty.Types.NetworkInterface

Description

Documentation

data NetworkInterface

Contains information about the elastic network interface of the EC2 instance.

See: newNetworkInterface smart constructor.

Constructors

NetworkInterface'

Fields

privateIpAddresses :: Maybe [PrivateIpAddressDetails]

Other private IP address information of the EC2 instance.

publicDnsName :: Maybe Text

The public DNS name of the EC2 instance.

securityGroups :: Maybe [SecurityGroup]

The security groups associated with the EC2 instance.

vpcId :: Maybe Text

The VPC ID of the EC2 instance.

networkInterfaceId :: Maybe Text

The ID of the network interface.

subnetId :: Maybe Text

The subnet ID of the EC2 instance.

privateIpAddress :: Maybe Text

The private IP address of the EC2 instance.

publicIp :: Maybe Text

The public IP address of the EC2 instance.

privateDnsName :: Maybe Text

The private DNS name of the EC2 instance.

ipv6Addresses :: Maybe [Text]

A list of IPv6 addresses for the EC2 instance.

newNetworkInterface :: NetworkInterface

Create a value of NetworkInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:privateIpAddresses:NetworkInterface', networkInterface_privateIpAddresses - Other private IP address information of the EC2 instance.

$sel:publicDnsName:NetworkInterface', networkInterface_publicDnsName - The public DNS name of the EC2 instance.

$sel:securityGroups:NetworkInterface', networkInterface_securityGroups - The security groups associated with the EC2 instance.

$sel:vpcId:NetworkInterface', networkInterface_vpcId - The VPC ID of the EC2 instance.

$sel:networkInterfaceId:NetworkInterface', networkInterface_networkInterfaceId - The ID of the network interface.

$sel:subnetId:NetworkInterface', networkInterface_subnetId - The subnet ID of the EC2 instance.

$sel:privateIpAddress:NetworkInterface', networkInterface_privateIpAddress - The private IP address of the EC2 instance.

$sel:publicIp:NetworkInterface', networkInterface_publicIp - The public IP address of the EC2 instance.

$sel:privateDnsName:NetworkInterface', networkInterface_privateDnsName - The private DNS name of the EC2 instance.

$sel:ipv6Addresses:NetworkInterface', networkInterface_ipv6Addresses - A list of IPv6 addresses for the EC2 instance.

networkInterface_privateIpAddresses :: Lens' NetworkInterface (Maybe [PrivateIpAddressDetails])

Other private IP address information of the EC2 instance.

networkInterface_publicDnsName :: Lens' NetworkInterface (Maybe Text)

The public DNS name of the EC2 instance.

networkInterface_securityGroups :: Lens' NetworkInterface (Maybe [SecurityGroup])

The security groups associated with the EC2 instance.

networkInterface_vpcId :: Lens' NetworkInterface (Maybe Text)

The VPC ID of the EC2 instance.

networkInterface_networkInterfaceId :: Lens' NetworkInterface (Maybe Text)

The ID of the network interface.

networkInterface_subnetId :: Lens' NetworkInterface (Maybe Text)

The subnet ID of the EC2 instance.

networkInterface_privateIpAddress :: Lens' NetworkInterface (Maybe Text)

The private IP address of the EC2 instance.

networkInterface_publicIp :: Lens' NetworkInterface (Maybe Text)

The public IP address of the EC2 instance.

networkInterface_privateDnsName :: Lens' NetworkInterface (Maybe Text)

The private DNS name of the EC2 instance.

networkInterface_ipv6Addresses :: Lens' NetworkInterface (Maybe [Text])

A list of IPv6 addresses for the EC2 instance.