Network.AWS.GuardDuty.Types.S3BucketDetail

Description

Documentation

data S3BucketDetail

Contains information on the S3 bucket.

See: newS3BucketDetail smart constructor.

Constructors

S3BucketDetail'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the S3 bucket.

createdAt :: Maybe POSIX

The date and time the bucket was created at.

owner :: Maybe Owner

The owner of the S3 bucket.

name :: Maybe Text

The name of the S3 bucket.

defaultServerSideEncryption :: Maybe DefaultServerSideEncryption

Describes the server side encryption method used in the S3 bucket.

publicAccess :: Maybe PublicAccess

Describes the public access policies that apply to the S3 bucket.

type' :: Maybe Text

Describes whether the bucket is a source or destination bucket.

tags :: Maybe [Tag]

All tags attached to the S3 bucket

newS3BucketDetail :: S3BucketDetail

Create a value of S3BucketDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:S3BucketDetail', s3BucketDetail_arn - The Amazon Resource Name (ARN) of the S3 bucket.

$sel:createdAt:S3BucketDetail', s3BucketDetail_createdAt - The date and time the bucket was created at.

$sel:owner:S3BucketDetail', s3BucketDetail_owner - The owner of the S3 bucket.

$sel:name:S3BucketDetail', s3BucketDetail_name - The name of the S3 bucket.

$sel:defaultServerSideEncryption:S3BucketDetail', s3BucketDetail_defaultServerSideEncryption - Describes the server side encryption method used in the S3 bucket.

$sel:publicAccess:S3BucketDetail', s3BucketDetail_publicAccess - Describes the public access policies that apply to the S3 bucket.

$sel:type':S3BucketDetail', s3BucketDetail_type - Describes whether the bucket is a source or destination bucket.

$sel:tags:S3BucketDetail', s3BucketDetail_tags - All tags attached to the S3 bucket

s3BucketDetail_arn :: Lens' S3BucketDetail (Maybe Text)

The Amazon Resource Name (ARN) of the S3 bucket.

s3BucketDetail_createdAt :: Lens' S3BucketDetail (Maybe UTCTime)

The date and time the bucket was created at.

s3BucketDetail_owner :: Lens' S3BucketDetail (Maybe Owner)

The owner of the S3 bucket.

s3BucketDetail_name :: Lens' S3BucketDetail (Maybe Text)

The name of the S3 bucket.

s3BucketDetail_defaultServerSideEncryption :: Lens' S3BucketDetail (Maybe DefaultServerSideEncryption)

Describes the server side encryption method used in the S3 bucket.

s3BucketDetail_publicAccess :: Lens' S3BucketDetail (Maybe PublicAccess)

Describes the public access policies that apply to the S3 bucket.

s3BucketDetail_type :: Lens' S3BucketDetail (Maybe Text)

Describes whether the bucket is a source or destination bucket.

s3BucketDetail_tags :: Lens' S3BucketDetail (Maybe [Tag])

All tags attached to the S3 bucket