Network.AWS.GuardDuty.UpdateDetector

Description

Updates the Amazon GuardDuty detector specified by the detectorId.

Creating a Request

data UpdateDetector

See: newUpdateDetector smart constructor.

Constructors

UpdateDetector'

Fields

findingPublishingFrequency :: Maybe FindingPublishingFrequency

An enum value that specifies how frequently findings are exported, such as to CloudWatch Events.

dataSources :: Maybe DataSourceConfigurations

Describes which data sources will be updated.

enable :: Maybe Bool

Specifies whether the detector is enabled or not enabled.

detectorId :: Text

The unique ID of the detector to update.

newUpdateDetector

Arguments

:: Text
$sel:detectorId:UpdateDetector'

-> UpdateDetector

Create a value of UpdateDetector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:findingPublishingFrequency:UpdateDetector', updateDetector_findingPublishingFrequency - An enum value that specifies how frequently findings are exported, such as to CloudWatch Events.

$sel:dataSources:UpdateDetector', updateDetector_dataSources - Describes which data sources will be updated.

$sel:enable:UpdateDetector', updateDetector_enable - Specifies whether the detector is enabled or not enabled.

$sel:detectorId:UpdateDetector', updateDetector_detectorId - The unique ID of the detector to update.

Request Lenses

updateDetector_findingPublishingFrequency :: Lens' UpdateDetector (Maybe FindingPublishingFrequency)

An enum value that specifies how frequently findings are exported, such as to CloudWatch Events.

updateDetector_dataSources :: Lens' UpdateDetector (Maybe DataSourceConfigurations)

Describes which data sources will be updated.

updateDetector_enable :: Lens' UpdateDetector (Maybe Bool)

Specifies whether the detector is enabled or not enabled.

updateDetector_detectorId :: Lens' UpdateDetector Text

The unique ID of the detector to update.

Destructuring the Response

data UpdateDetectorResponse

See: newUpdateDetectorResponse smart constructor.

Constructors

UpdateDetectorResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateDetectorResponse

Arguments

:: Int
$sel:httpStatus:UpdateDetectorResponse'

-> UpdateDetectorResponse

Create a value of UpdateDetectorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateDetectorResponse', updateDetectorResponse_httpStatus - The response's http status code.

Response Lenses

updateDetectorResponse_httpStatus :: Lens' UpdateDetectorResponse Int

The response's http status code.