Network.AWS.GuardDuty.UpdatePublishingDestination

Description

Updates information about the publishing destination specified by the destinationId.

Creating a Request

data UpdatePublishingDestination

See: newUpdatePublishingDestination smart constructor.

Constructors

UpdatePublishingDestination'

Fields

destinationProperties :: Maybe DestinationProperties

A DestinationProperties object that includes the DestinationArn and KmsKeyArn of the publishing destination.

detectorId :: Text

The ID of the detector associated with the publishing destinations to update.

destinationId :: Text

The ID of the publishing destination to update.

newUpdatePublishingDestination

Arguments

:: Text
$sel:detectorId:UpdatePublishingDestination'

-> Text
$sel:destinationId:UpdatePublishingDestination'

-> UpdatePublishingDestination

Create a value of UpdatePublishingDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationProperties:UpdatePublishingDestination', updatePublishingDestination_destinationProperties - A DestinationProperties object that includes the DestinationArn and KmsKeyArn of the publishing destination.

$sel:detectorId:UpdatePublishingDestination', updatePublishingDestination_detectorId - The ID of the detector associated with the publishing destinations to update.

$sel:destinationId:UpdatePublishingDestination', updatePublishingDestination_destinationId - The ID of the publishing destination to update.

Request Lenses

updatePublishingDestination_destinationProperties :: Lens' UpdatePublishingDestination (Maybe DestinationProperties)

A DestinationProperties object that includes the DestinationArn and KmsKeyArn of the publishing destination.

updatePublishingDestination_detectorId :: Lens' UpdatePublishingDestination Text

The ID of the detector associated with the publishing destinations to update.

updatePublishingDestination_destinationId :: Lens' UpdatePublishingDestination Text

The ID of the publishing destination to update.

Destructuring the Response

data UpdatePublishingDestinationResponse

See: newUpdatePublishingDestinationResponse smart constructor.

Constructors

UpdatePublishingDestinationResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdatePublishingDestinationResponse

Arguments

:: Int
$sel:httpStatus:UpdatePublishingDestinationResponse'

-> UpdatePublishingDestinationResponse

Create a value of UpdatePublishingDestinationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdatePublishingDestinationResponse', updatePublishingDestinationResponse_httpStatus - The response's http status code.

Response Lenses

updatePublishingDestinationResponse_httpStatus :: Lens' UpdatePublishingDestinationResponse Int

The response's http status code.