Network.AWS.Headers

Documentation

flattenHeaders :: [AnyHeader] -> [(ByteString, ByteString)]

lookupHeader :: ByteString -> [AnyHeader] -> Maybe ByteString

class IsHeader a where

Methods

encodeHeader :: a -> ByteString -> (ByteString, ByteString)

data Header k v

Constructors

Header v

data AnyHeader where

Constructors

AnyHeader :: IsHeader a => a -> AnyHeader

hdr :: IsHeader a => a -> AnyHeader

type ContentLength = Header "content-length" Integer

type ContentLanguage = Header "content-language" ByteString

type Expect = Header "expect" ByteString

type Expires = Header "expires" ByteString

type Range = Header "range" ByteString

type IfModifiedSince = Header "if-modified-since" ByteString

type IfUnmodifiedSince = Header "if-unmodified-since" ByteString

type IfMatch = Header "if-match" ByteString

type IfNoneMatch = Header "if-none-match" ByteString

data Content t s

Constructors

Content

contentType :: SingI t => Content t s -> ByteString

contentSubType :: SingI s => Content t s -> ByteString

type JSON = Content "application" "json"

type XML = Content "application" "xml"

type FormURLEncoded = Content "application" "x-www-form-urlencoded"

class CacheValue a where

Methods

cacheValue :: ByteString -> a -> ByteString

newtype Cache k v

Constructors

Cache v

type Public = Cache "public" ()

type Private = Cache "private" (Maybe Text)

type NoCache = Cache "no-cache" (Maybe Text)

type NoStore = Cache "no-store" ()

type NoTransform = Cache "no-transform" ()

type MustRevalidate = Cache "must-revalidate" ()

type ProxyRevalidate = Cache "proxy-revalidate" ()

type MaxAge = Cache "max-age" Integer

type SMaxAge = Cache "s-maxage" Integer

type MaxStale = Cache "max-stale" (Maybe Integer)

type MinFresh = Cache "min-fresh" Integer

type OnlyIfCache = Cache "only-if-cached" ()

data Encoding t

Constructors

Encoding

type GZip = Encoding "gzip"

type Deflate = Encoding "deflate"

newtype Param k v

Constructors

Param v

class ParamValue a where

Methods

paramValue :: a -> ByteString

data AnyParam where

Constructors

AnyParam :: ParamValue a => a -> AnyParam

prm :: ParamValue a => a -> AnyParam

newtype Disposition t

Constructors

Disposition [AnyParam]

type Inline = Disposition "inline"

Displayed automatically [RFC2183]

type Attachment = Disposition "attachment"

Attachment user controlled display [RFC2183]

type FormData = Disposition "form-data"

Process as form response [RFC2388]

type Signal = Disposition "signal"

Tunneled content to be processed silently [RFC3204]

type Alert = Disposition "alert"

Custom ring tone to alert the user [RFC3261]

type Icon = Disposition "icon"

Displayed as an icon to the user [RFC3261]

type Render = Disposition "render"

Should be displayed to the user [RFC3261]

type RecipientListHistory = Disposition "recipient-list-history"

Contains a list of URIs that indicates the recipients of the request [RFC5364]

type Session = Disposition "session"

Describes a communications session. For example, an RFC2327 SDP body [RFC3261]

type AIB = Disposition "aib"

Authenticated Identity Body [RFC3893]

type EarlySession = Disposition "early-session"

Describes an early communications session. For example, and [RFC2327] SDP body [RFC3959]

type RecipientList = Disposition "recipient-list"

Includes a list of URIs to which URI-list services are to be applied. [RFC5363]

type Notification = Disposition "notification"

Payload of the message carrying this Content-Disposition header field value is an Instant Message Disposition Notification as requested in the corresponding Instant Message. [RFC5438]

type ByReference = Disposition "by-reference"

Needs to be handled according to a reference to the body that is located in the same SIP message as the body. [RFC5621]

type InfoPackage = Disposition "info-package"

Contains information associated with an Info Package

type FileName = Param "filename" ByteString

Name to be used when creating file [RFC2183]

type CreationDate = Param "creation-date" UTCTime

When content was created [RFC2183]

type ModificationDate = Param "modification-date" UTCTime

When content was last modified [RFC2183]

type ReadDate = Param "read-date" UTCTime

When content was last read [RFC2183]

type Size = Param "size" Integer

Approximate size of content in octets [RFC2183]

type Name = Param "name" ByteString

Original field name in form [RFC2388]

type Handling = Param "handling" HandlingType

Whether or not processing is required [RFC3204]

type Voice = Param "voice" VoiceType

Type or use of audio content [RFC2421]

data HandlingType

Constructors

Required

Optional

data VoiceType

Constructors

VoiceMessage

VoiceMessageNotification

OriginatorSpokenName

RecipientSpokenName

SpokenSubject

newtype MD5

Constructors

MD5 ByteString

md5 :: ByteString -> MD5

class ToHeaders a where

Supplementary class to extract any header types from a record, ignoring everything else.

Methods

toHeaders :: a -> [AnyHeader]

genericHeaders :: (Generic a, GHeaders (Rep a)) => a -> [AnyHeader]

class GHeaders f where

Methods

gHeaders :: f a -> [AnyHeader]