Network.AWS.HealthLake

Description

Derived from API version 2017-07-01 of the AWS service descriptions, licensed under Apache 2.0.

Amazon HealthLake is a HIPAA eligibile service that allows customers to store, transform, query, and analyze their FHIR-formatted data in a consistent fashion in the cloud.

Service Configuration

defaultService :: Service

API version 2017-07-01 of the Amazon HealthLake SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by HealthLake.

ValidationException

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The user input parameter was invalid.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

Access is denied. Your account is not authorized to perform this operation.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The Data Store is in a transition state and the user requested action can not be performed.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The user has exceeded their maximum number of allowed calls to the given API.

InternalServerException

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

Unknown error occurs in the service.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The requested Data Store was not found.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

StartFHIRImportJob

data StartFHIRImportJob

See: newStartFHIRImportJob smart constructor.

Constructors

StartFHIRImportJob' (Maybe Text) InputDataConfig OutputDataConfig Text Text Text

newStartFHIRImportJob

Arguments

:: InputDataConfig
$sel:inputDataConfig:StartFHIRImportJob'

-> OutputDataConfig
$sel:jobOutputDataConfig:StartFHIRImportJob'

-> Text
$sel:datastoreId:StartFHIRImportJob'

-> Text
$sel:dataAccessRoleArn:StartFHIRImportJob'

-> Text
$sel:clientToken:StartFHIRImportJob'

-> StartFHIRImportJob

Create a value of StartFHIRImportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobName:StartFHIRImportJob', startFHIRImportJob_jobName - The name of the FHIR Import job in the StartFHIRImport job request.

$sel:inputDataConfig:StartFHIRImportJob', startFHIRImportJob_inputDataConfig - The input properties of the FHIR Import job in the StartFHIRImport job request.

$sel:jobOutputDataConfig:StartFHIRImportJob', startFHIRImportJob_jobOutputDataConfig - Undocumented member.

$sel:datastoreId:StartFHIRImportJob', startFHIRImportJob_datastoreId - The AWS-generated Data Store ID.

$sel:dataAccessRoleArn:StartFHIRImportJob', startFHIRImportJob_dataAccessRoleArn - The Amazon Resource Name (ARN) that gives Amazon HealthLake access permission.

$sel:clientToken:StartFHIRImportJob', startFHIRImportJob_clientToken - Optional user provided token used for ensuring idempotency.

data StartFHIRImportJobResponse

See: newStartFHIRImportJobResponse smart constructor.

Constructors

StartFHIRImportJobResponse' (Maybe Text) Int Text JobStatus

newStartFHIRImportJobResponse

Arguments

:: Int
$sel:httpStatus:StartFHIRImportJobResponse'

-> Text
$sel:jobId:StartFHIRImportJobResponse'

-> JobStatus
$sel:jobStatus:StartFHIRImportJobResponse'

-> StartFHIRImportJobResponse

Create a value of StartFHIRImportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datastoreId:StartFHIRImportJob', startFHIRImportJobResponse_datastoreId - The AWS-generated Data Store ID.

$sel:httpStatus:StartFHIRImportJobResponse', startFHIRImportJobResponse_httpStatus - The response's http status code.

$sel:jobId:StartFHIRImportJobResponse', startFHIRImportJobResponse_jobId - The AWS-generated job ID.

$sel:jobStatus:StartFHIRImportJobResponse', startFHIRImportJobResponse_jobStatus - The status of an import job.

DescribeFHIRDatastore

data DescribeFHIRDatastore

See: newDescribeFHIRDatastore smart constructor.

Constructors

DescribeFHIRDatastore' (Maybe Text)

newDescribeFHIRDatastore :: DescribeFHIRDatastore

Create a value of DescribeFHIRDatastore with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datastoreId:DescribeFHIRDatastore', describeFHIRDatastore_datastoreId - The AWS-generated Data Store id. This is part of the ‘CreateFHIRDatastore’ output.

data DescribeFHIRDatastoreResponse

See: newDescribeFHIRDatastoreResponse smart constructor.

Constructors

DescribeFHIRDatastoreResponse' Int DatastoreProperties

newDescribeFHIRDatastoreResponse

Arguments

:: Int
$sel:httpStatus:DescribeFHIRDatastoreResponse'

-> DatastoreProperties
$sel:datastoreProperties:DescribeFHIRDatastoreResponse'

-> DescribeFHIRDatastoreResponse

Create a value of DescribeFHIRDatastoreResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeFHIRDatastoreResponse', describeFHIRDatastoreResponse_httpStatus - The response's http status code.

$sel:datastoreProperties:DescribeFHIRDatastoreResponse', describeFHIRDatastoreResponse_datastoreProperties - All properties associated with a Data Store, including the Data Store ID, Data Store ARN, Data Store name, Data Store status, created at, Data Store type version, and Data Store endpoint.

DescribeFHIRImportJob

data DescribeFHIRImportJob

See: newDescribeFHIRImportJob smart constructor.

Constructors

DescribeFHIRImportJob' Text Text

newDescribeFHIRImportJob

Arguments

:: Text
$sel:datastoreId:DescribeFHIRImportJob'

-> Text
$sel:jobId:DescribeFHIRImportJob'

-> DescribeFHIRImportJob

Create a value of DescribeFHIRImportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datastoreId:DescribeFHIRImportJob', describeFHIRImportJob_datastoreId - The AWS-generated ID of the Data Store.

$sel:jobId:DescribeFHIRImportJob', describeFHIRImportJob_jobId - The AWS-generated job ID.

data DescribeFHIRImportJobResponse

See: newDescribeFHIRImportJobResponse smart constructor.

Constructors

DescribeFHIRImportJobResponse' Int ImportJobProperties

newDescribeFHIRImportJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeFHIRImportJobResponse'

-> ImportJobProperties
$sel:importJobProperties:DescribeFHIRImportJobResponse'

-> DescribeFHIRImportJobResponse

Create a value of DescribeFHIRImportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeFHIRImportJobResponse', describeFHIRImportJobResponse_httpStatus - The response's http status code.

$sel:importJobProperties:DescribeFHIRImportJobResponse', describeFHIRImportJobResponse_importJobProperties - The properties of the Import job request, including the ID, ARN, name, and the status of the job.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceARN:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceARN:ListTagsForResource', listTagsForResource_resourceARN - The Amazon Resource Name(ARN) of the Data Store for which tags are being added.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe [Tag]) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - Returns a list of tags associated with a Data Store.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

DeleteFHIRDatastore

data DeleteFHIRDatastore

See: newDeleteFHIRDatastore smart constructor.

Constructors

DeleteFHIRDatastore' (Maybe Text)

newDeleteFHIRDatastore :: DeleteFHIRDatastore

Create a value of DeleteFHIRDatastore with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datastoreId:DeleteFHIRDatastore', deleteFHIRDatastore_datastoreId - The AWS-generated ID for the Data Store to be deleted.

data DeleteFHIRDatastoreResponse

See: newDeleteFHIRDatastoreResponse smart constructor.

Constructors

DeleteFHIRDatastoreResponse' Int Text Text DatastoreStatus Text

newDeleteFHIRDatastoreResponse

Arguments

:: Int
$sel:httpStatus:DeleteFHIRDatastoreResponse'

-> Text
$sel:datastoreId:DeleteFHIRDatastore'

-> Text
$sel:datastoreArn:DeleteFHIRDatastoreResponse'

-> DatastoreStatus
$sel:datastoreStatus:DeleteFHIRDatastoreResponse'

-> Text
$sel:datastoreEndpoint:DeleteFHIRDatastoreResponse'

-> DeleteFHIRDatastoreResponse

Create a value of DeleteFHIRDatastoreResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteFHIRDatastoreResponse', deleteFHIRDatastoreResponse_httpStatus - The response's http status code.

$sel:datastoreId:DeleteFHIRDatastore', deleteFHIRDatastoreResponse_datastoreId - The AWS-generated ID for the Data Store to be deleted.

$sel:datastoreArn:DeleteFHIRDatastoreResponse', deleteFHIRDatastoreResponse_datastoreArn - The Amazon Resource Name (ARN) that gives Amazon HealthLake access permission.

$sel:datastoreStatus:DeleteFHIRDatastoreResponse', deleteFHIRDatastoreResponse_datastoreStatus - The status of the Data Store that the user has requested to be deleted.

$sel:datastoreEndpoint:DeleteFHIRDatastoreResponse', deleteFHIRDatastoreResponse_datastoreEndpoint - The AWS endpoint for the Data Store the user has requested to be deleted.

DescribeFHIRExportJob

data DescribeFHIRExportJob

See: newDescribeFHIRExportJob smart constructor.

Constructors

DescribeFHIRExportJob' Text Text

newDescribeFHIRExportJob

Arguments

:: Text
$sel:datastoreId:DescribeFHIRExportJob'

-> Text
$sel:jobId:DescribeFHIRExportJob'

-> DescribeFHIRExportJob

Create a value of DescribeFHIRExportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datastoreId:DescribeFHIRExportJob', describeFHIRExportJob_datastoreId - The AWS generated ID for the Data Store from which files are being exported from for an export job.

$sel:jobId:DescribeFHIRExportJob', describeFHIRExportJob_jobId - The AWS generated ID for an export job.

data DescribeFHIRExportJobResponse

See: newDescribeFHIRExportJobResponse smart constructor.

Constructors

DescribeFHIRExportJobResponse' Int ExportJobProperties

newDescribeFHIRExportJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeFHIRExportJobResponse'

-> ExportJobProperties
$sel:exportJobProperties:DescribeFHIRExportJobResponse'

-> DescribeFHIRExportJobResponse

Create a value of DescribeFHIRExportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeFHIRExportJobResponse', describeFHIRExportJobResponse_httpStatus - The response's http status code.

$sel:exportJobProperties:DescribeFHIRExportJobResponse', describeFHIRExportJobResponse_exportJobProperties - Displays the properties of the export job, including the ID, Arn, Name, and the status of the job.

CreateFHIRDatastore

data CreateFHIRDatastore

See: newCreateFHIRDatastore smart constructor.

Constructors

CreateFHIRDatastore' (Maybe Text) (Maybe SseConfiguration) (Maybe Text) (Maybe PreloadDataConfig) (Maybe [Tag]) FHIRVersion

newCreateFHIRDatastore

Arguments

:: FHIRVersion
$sel:datastoreTypeVersion:CreateFHIRDatastore'

-> CreateFHIRDatastore

Create a value of CreateFHIRDatastore with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateFHIRDatastore', createFHIRDatastore_clientToken - Optional user provided token used for ensuring idempotency.

$sel:sseConfiguration:CreateFHIRDatastore', createFHIRDatastore_sseConfiguration - The server-side encryption key configuration for a customer provided encryption key specified for creating a Data Store.

$sel:datastoreName:CreateFHIRDatastore', createFHIRDatastore_datastoreName - The user generated name for the Data Store.

$sel:preloadDataConfig:CreateFHIRDatastore', createFHIRDatastore_preloadDataConfig - Optional parameter to preload data upon creation of the Data Store. Currently, the only supported preloaded data is synthetic data generated from Synthea.

$sel:tags:CreateFHIRDatastore', createFHIRDatastore_tags - Resource tags that are applied to a Data Store when it is created.

$sel:datastoreTypeVersion:CreateFHIRDatastore', createFHIRDatastore_datastoreTypeVersion - The FHIR version of the Data Store. The only supported version is R4.

data CreateFHIRDatastoreResponse

See: newCreateFHIRDatastoreResponse smart constructor.

Constructors

CreateFHIRDatastoreResponse' Int Text Text DatastoreStatus Text

newCreateFHIRDatastoreResponse

Arguments

:: Int
$sel:httpStatus:CreateFHIRDatastoreResponse'

-> Text
$sel:datastoreId:CreateFHIRDatastoreResponse'

-> Text
$sel:datastoreArn:CreateFHIRDatastoreResponse'

-> DatastoreStatus
$sel:datastoreStatus:CreateFHIRDatastoreResponse'

-> Text
$sel:datastoreEndpoint:CreateFHIRDatastoreResponse'

-> CreateFHIRDatastoreResponse

Create a value of CreateFHIRDatastoreResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateFHIRDatastoreResponse', createFHIRDatastoreResponse_httpStatus - The response's http status code.

$sel:datastoreId:CreateFHIRDatastoreResponse', createFHIRDatastoreResponse_datastoreId - The AWS-generated Data Store id. This id is in the output from the initial Data Store creation call.

$sel:datastoreArn:CreateFHIRDatastoreResponse', createFHIRDatastoreResponse_datastoreArn - The datastore ARN is generated during the creation of the Data Store and can be found in the output from the initial Data Store creation call.

$sel:datastoreStatus:CreateFHIRDatastoreResponse', createFHIRDatastoreResponse_datastoreStatus - The status of the FHIR Data Store. Possible statuses are ‘CREATING’, ‘ACTIVE’, ‘DELETING’, ‘DELETED’.

$sel:datastoreEndpoint:CreateFHIRDatastoreResponse', createFHIRDatastoreResponse_datastoreEndpoint - The AWS endpoint for the created Data Store. For preview, only US-east-1 endpoints are supported.

ListFHIRExportJobs

data ListFHIRExportJobs

See: newListFHIRExportJobs smart constructor.

Constructors

ListFHIRExportJobs' (Maybe POSIX) (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe JobStatus) (Maybe Natural) Text

newListFHIRExportJobs

Arguments

:: Text
$sel:datastoreId:ListFHIRExportJobs'

-> ListFHIRExportJobs

Create a value of ListFHIRExportJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:submittedAfter:ListFHIRExportJobs', listFHIRExportJobs_submittedAfter - This parameter limits the response to FHIR export jobs submitted after a user specified date.

$sel:jobName:ListFHIRExportJobs', listFHIRExportJobs_jobName - This parameter limits the response to the export job with the specified job name.

$sel:submittedBefore:ListFHIRExportJobs', listFHIRExportJobs_submittedBefore - This parameter limits the response to FHIR export jobs submitted before a user specified date.

$sel:nextToken:ListFHIRExportJobs', listFHIRExportJobs_nextToken - A pagination token used to identify the next page of results to return for a ListFHIRExportJobs query.

$sel:jobStatus:ListFHIRExportJobs', listFHIRExportJobs_jobStatus - This parameter limits the response to the export jobs with the specified job status.

$sel:maxResults:ListFHIRExportJobs', listFHIRExportJobs_maxResults - This parameter limits the number of results returned for a ListFHIRExportJobs to a maximum quantity specified by the user.

$sel:datastoreId:ListFHIRExportJobs', listFHIRExportJobs_datastoreId - This parameter limits the response to the export job with the specified Data Store ID.

data ListFHIRExportJobsResponse

See: newListFHIRExportJobsResponse smart constructor.

Constructors

ListFHIRExportJobsResponse' (Maybe Text) Int [ExportJobProperties]

newListFHIRExportJobsResponse

Arguments

:: Int
$sel:httpStatus:ListFHIRExportJobsResponse'

-> ListFHIRExportJobsResponse

Create a value of ListFHIRExportJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFHIRExportJobs', listFHIRExportJobsResponse_nextToken - A pagination token used to identify the next page of results to return for a ListFHIRExportJobs query.

$sel:httpStatus:ListFHIRExportJobsResponse', listFHIRExportJobsResponse_httpStatus - The response's http status code.

$sel:exportJobPropertiesList:ListFHIRExportJobsResponse', listFHIRExportJobsResponse_exportJobPropertiesList - The properties of listed FHIR export jobs, including the ID, ARN, name, and the status of the job.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text [Tag]

newTagResource

Arguments

:: Text
$sel:resourceARN:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceARN:TagResource', tagResource_resourceARN - The Amazon Resource Name(ARN)that gives Amazon HealthLake access to the Data Store which tags are being added to.

$sel:tags:TagResource', tagResource_tags - The user specified key and value pair tags being added to a Data Store.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text [Text]

newUntagResource

Arguments

:: Text
$sel:resourceARN:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceARN:UntagResource', untagResource_resourceARN - "The Amazon Resource Name(ARN) of the Data Store for which tags are being removed

$sel:tagKeys:UntagResource', untagResource_tagKeys - The keys for the tags to be removed from the Healthlake Data Store.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

ListFHIRDatastores

data ListFHIRDatastores

See: newListFHIRDatastores smart constructor.

Constructors

ListFHIRDatastores' (Maybe Text) (Maybe DatastoreFilter) (Maybe Natural)

newListFHIRDatastores :: ListFHIRDatastores

Create a value of ListFHIRDatastores with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFHIRDatastores', listFHIRDatastores_nextToken - Fetches the next page of Data Stores when results are paginated.

$sel:filter':ListFHIRDatastores', listFHIRDatastores_filter - Lists all filters associated with a FHIR Data Store request.

$sel:maxResults:ListFHIRDatastores', listFHIRDatastores_maxResults - The maximum number of Data Stores returned in a single page of a ListFHIRDatastoresRequest call.

data ListFHIRDatastoresResponse

See: newListFHIRDatastoresResponse smart constructor.

Constructors

ListFHIRDatastoresResponse' (Maybe Text) Int [DatastoreProperties]

newListFHIRDatastoresResponse

Arguments

:: Int
$sel:httpStatus:ListFHIRDatastoresResponse'

-> ListFHIRDatastoresResponse

Create a value of ListFHIRDatastoresResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFHIRDatastores', listFHIRDatastoresResponse_nextToken - Pagination token that can be used to retrieve the next page of results.

$sel:httpStatus:ListFHIRDatastoresResponse', listFHIRDatastoresResponse_httpStatus - The response's http status code.

$sel:datastorePropertiesList:ListFHIRDatastoresResponse', listFHIRDatastoresResponse_datastorePropertiesList - All properties associated with the listed Data Stores.

StartFHIRExportJob

data StartFHIRExportJob

See: newStartFHIRExportJob smart constructor.

Constructors

StartFHIRExportJob' (Maybe Text) OutputDataConfig Text Text Text

newStartFHIRExportJob

Arguments

:: OutputDataConfig
$sel:outputDataConfig:StartFHIRExportJob'

-> Text
$sel:datastoreId:StartFHIRExportJob'

-> Text
$sel:dataAccessRoleArn:StartFHIRExportJob'

-> Text
$sel:clientToken:StartFHIRExportJob'

-> StartFHIRExportJob

Create a value of StartFHIRExportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobName:StartFHIRExportJob', startFHIRExportJob_jobName - The user generated name for an export job.

$sel:outputDataConfig:StartFHIRExportJob', startFHIRExportJob_outputDataConfig - The output data configuration that was supplied when the export job was created.

$sel:datastoreId:StartFHIRExportJob', startFHIRExportJob_datastoreId - The AWS generated ID for the Data Store from which files are being exported for an export job.

$sel:dataAccessRoleArn:StartFHIRExportJob', startFHIRExportJob_dataAccessRoleArn - The Amazon Resource Name used during the initiation of the job.

$sel:clientToken:StartFHIRExportJob', startFHIRExportJob_clientToken - An optional user provided token used for ensuring idempotency.

data StartFHIRExportJobResponse

See: newStartFHIRExportJobResponse smart constructor.

Constructors

StartFHIRExportJobResponse' (Maybe Text) Int Text JobStatus

newStartFHIRExportJobResponse

Arguments

:: Int
$sel:httpStatus:StartFHIRExportJobResponse'

-> Text
$sel:jobId:StartFHIRExportJobResponse'

-> JobStatus
$sel:jobStatus:StartFHIRExportJobResponse'

-> StartFHIRExportJobResponse

Create a value of StartFHIRExportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datastoreId:StartFHIRExportJob', startFHIRExportJobResponse_datastoreId - The AWS generated ID for the Data Store from which files are being exported for an export job.

$sel:httpStatus:StartFHIRExportJobResponse', startFHIRExportJobResponse_httpStatus - The response's http status code.

$sel:jobId:StartFHIRExportJobResponse', startFHIRExportJobResponse_jobId - The AWS generated ID for an export job.

$sel:jobStatus:StartFHIRExportJobResponse', startFHIRExportJobResponse_jobStatus - The status of a FHIR export job. Possible statuses are SUBMITTED, IN_PROGRESS, COMPLETED, or FAILED.

ListFHIRImportJobs

data ListFHIRImportJobs

See: newListFHIRImportJobs smart constructor.

Constructors

ListFHIRImportJobs' (Maybe POSIX) (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe JobStatus) (Maybe Natural) Text

newListFHIRImportJobs

Arguments

:: Text
$sel:datastoreId:ListFHIRImportJobs'

-> ListFHIRImportJobs

Create a value of ListFHIRImportJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:submittedAfter:ListFHIRImportJobs', listFHIRImportJobs_submittedAfter - This parameter limits the response to FHIR import jobs submitted after a user specified date.

$sel:jobName:ListFHIRImportJobs', listFHIRImportJobs_jobName - This parameter limits the response to the import job with the specified job name.

$sel:submittedBefore:ListFHIRImportJobs', listFHIRImportJobs_submittedBefore - This parameter limits the response to FHIR import jobs submitted before a user specified date.

$sel:nextToken:ListFHIRImportJobs', listFHIRImportJobs_nextToken - A pagination token used to identify the next page of results to return for a ListFHIRImportJobs query.

$sel:jobStatus:ListFHIRImportJobs', listFHIRImportJobs_jobStatus - This parameter limits the response to the import job with the specified job status.

$sel:maxResults:ListFHIRImportJobs', listFHIRImportJobs_maxResults - This parameter limits the number of results returned for a ListFHIRImportJobs to a maximum quantity specified by the user.

$sel:datastoreId:ListFHIRImportJobs', listFHIRImportJobs_datastoreId - This parameter limits the response to the import job with the specified Data Store ID.

data ListFHIRImportJobsResponse

See: newListFHIRImportJobsResponse smart constructor.

Constructors

ListFHIRImportJobsResponse' (Maybe Text) Int [ImportJobProperties]

newListFHIRImportJobsResponse

Arguments

:: Int
$sel:httpStatus:ListFHIRImportJobsResponse'

-> ListFHIRImportJobsResponse

Create a value of ListFHIRImportJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFHIRImportJobs', listFHIRImportJobsResponse_nextToken - A pagination token used to identify the next page of results to return for a ListFHIRImportJobs query.

$sel:httpStatus:ListFHIRImportJobsResponse', listFHIRImportJobsResponse_httpStatus - The response's http status code.

$sel:importJobPropertiesList:ListFHIRImportJobsResponse', listFHIRImportJobsResponse_importJobPropertiesList - The properties of a listed FHIR import jobs, including the ID, ARN, name, and the status of the job.

Types

CmkType

newtype CmkType

Constructors

CmkType'

Fields

fromCmkType :: Text

Bundled Patterns

pattern CmkType_AWS_OWNED_KMS_KEY :: CmkType

pattern CmkType_CUSTOMER_MANAGED_KMS_KEY :: CmkType

DatastoreStatus

newtype DatastoreStatus

Constructors

DatastoreStatus'

Fields

fromDatastoreStatus :: Text

Bundled Patterns

pattern DatastoreStatus_ACTIVE :: DatastoreStatus

pattern DatastoreStatus_CREATING :: DatastoreStatus

pattern DatastoreStatus_DELETED :: DatastoreStatus

pattern DatastoreStatus_DELETING :: DatastoreStatus

FHIRVersion

newtype FHIRVersion

Constructors

FHIRVersion'

Fields

fromFHIRVersion :: Text

Bundled Patterns

pattern FHIRVersion_R4 :: FHIRVersion

JobStatus

newtype JobStatus

Constructors

JobStatus'

Fields

fromJobStatus :: Text

Bundled Patterns

pattern JobStatus_COMPLETED :: JobStatus

pattern JobStatus_COMPLETED_WITH_ERRORS :: JobStatus

pattern JobStatus_FAILED :: JobStatus

pattern JobStatus_IN_PROGRESS :: JobStatus

pattern JobStatus_SUBMITTED :: JobStatus

PreloadDataType

newtype PreloadDataType

Constructors

PreloadDataType'

Fields

fromPreloadDataType :: Text

Bundled Patterns

pattern PreloadDataType_SYNTHEA :: PreloadDataType

DatastoreFilter

data DatastoreFilter

The filters applied to Data Store query.

See: newDatastoreFilter smart constructor.

Constructors

DatastoreFilter' (Maybe POSIX) (Maybe Text) (Maybe DatastoreStatus) (Maybe POSIX)

newDatastoreFilter :: DatastoreFilter

Create a value of DatastoreFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAfter:DatastoreFilter', datastoreFilter_createdAfter - A filter that allows the user to set cutoff dates for records. All Data Stores created after the specified date will be included in the results.

$sel:datastoreName:DatastoreFilter', datastoreFilter_datastoreName - Allows the user to filter Data Store results by name.

$sel:datastoreStatus:DatastoreFilter', datastoreFilter_datastoreStatus - Allows the user to filter Data Store results by status.

$sel:createdBefore:DatastoreFilter', datastoreFilter_createdBefore - A filter that allows the user to set cutoff dates for records. All Data Stores created before the specified date will be included in the results.

DatastoreProperties

data DatastoreProperties

Displays the properties of the Data Store, including the ID, Arn, name, and the status of the Data Store.

See: newDatastoreProperties smart constructor.

Constructors

DatastoreProperties' (Maybe SseConfiguration) (Maybe POSIX) (Maybe Text) (Maybe PreloadDataConfig) Text Text DatastoreStatus FHIRVersion Text

newDatastoreProperties

Arguments

:: Text
$sel:datastoreId:DatastoreProperties'

-> Text
$sel:datastoreArn:DatastoreProperties'

-> DatastoreStatus
$sel:datastoreStatus:DatastoreProperties'

-> FHIRVersion
$sel:datastoreTypeVersion:DatastoreProperties'

-> Text
$sel:datastoreEndpoint:DatastoreProperties'

-> DatastoreProperties

Create a value of DatastoreProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sseConfiguration:DatastoreProperties', datastoreProperties_sseConfiguration - The server-side encryption key configuration for a customer provided encryption key (CMK).

$sel:createdAt:DatastoreProperties', datastoreProperties_createdAt - The time that a Data Store was created.

$sel:datastoreName:DatastoreProperties', datastoreProperties_datastoreName - The user-generated name for the Data Store.

$sel:preloadDataConfig:DatastoreProperties', datastoreProperties_preloadDataConfig - The preloaded data configuration for the Data Store. Only data preloaded from Synthea is supported.

$sel:datastoreId:DatastoreProperties', datastoreProperties_datastoreId - The AWS-generated ID number for the Data Store.

$sel:datastoreArn:DatastoreProperties', datastoreProperties_datastoreArn - The Amazon Resource Name used in the creation of the Data Store.

$sel:datastoreStatus:DatastoreProperties', datastoreProperties_datastoreStatus - The status of the Data Store. Possible statuses are 'CREATING', 'ACTIVE', 'DELETING', or 'DELETED'.

$sel:datastoreTypeVersion:DatastoreProperties', datastoreProperties_datastoreTypeVersion - The FHIR version. Only R4 version data is supported.

$sel:datastoreEndpoint:DatastoreProperties', datastoreProperties_datastoreEndpoint - The AWS endpoint for the Data Store. Each Data Store will have it's own endpoint with Data Store ID in the endpoint URL.

ExportJobProperties

data ExportJobProperties

The properties of a FHIR export job, including the ID, ARN, name, and the status of the job.

See: newExportJobProperties smart constructor.

Constructors

ExportJobProperties' (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe Text) Text JobStatus POSIX Text OutputDataConfig

newExportJobProperties

Arguments

:: Text
$sel:jobId:ExportJobProperties'

-> JobStatus
$sel:jobStatus:ExportJobProperties'

-> UTCTime
$sel:submitTime:ExportJobProperties'

-> Text
$sel:datastoreId:ExportJobProperties'

-> OutputDataConfig
$sel:outputDataConfig:ExportJobProperties'

-> ExportJobProperties

Create a value of ExportJobProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobName:ExportJobProperties', exportJobProperties_jobName - The user generated name for an export job.

$sel:endTime:ExportJobProperties', exportJobProperties_endTime - The time an export job completed.

$sel:dataAccessRoleArn:ExportJobProperties', exportJobProperties_dataAccessRoleArn - The Amazon Resource Name used during the initiation of the job.

$sel:message:ExportJobProperties', exportJobProperties_message - An explanation of any errors that may have occurred during the export job.

$sel:jobId:ExportJobProperties', exportJobProperties_jobId - The AWS generated ID for an export job.

$sel:jobStatus:ExportJobProperties', exportJobProperties_jobStatus - The status of a FHIR export job. Possible statuses are SUBMITTED, IN_PROGRESS, COMPLETED, or FAILED.

$sel:submitTime:ExportJobProperties', exportJobProperties_submitTime - The time an export job was initiated.

$sel:datastoreId:ExportJobProperties', exportJobProperties_datastoreId - The AWS generated ID for the Data Store from which files are being exported for an export job.

$sel:outputDataConfig:ExportJobProperties', exportJobProperties_outputDataConfig - The output data configuration that was supplied when the export job was created.

ImportJobProperties

data ImportJobProperties

Displays the properties of the import job, including the ID, Arn, Name, and the status of the Data Store.

See: newImportJobProperties smart constructor.

Constructors

ImportJobProperties' (Maybe OutputDataConfig) (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe Text) Text JobStatus POSIX Text InputDataConfig

newImportJobProperties

Arguments

:: Text
$sel:jobId:ImportJobProperties'

-> JobStatus
$sel:jobStatus:ImportJobProperties'

-> UTCTime
$sel:submitTime:ImportJobProperties'

-> Text
$sel:datastoreId:ImportJobProperties'

-> InputDataConfig
$sel:inputDataConfig:ImportJobProperties'

-> ImportJobProperties

Create a value of ImportJobProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobOutputDataConfig:ImportJobProperties', importJobProperties_jobOutputDataConfig - Undocumented member.

$sel:jobName:ImportJobProperties', importJobProperties_jobName - The user-generated name for an Import job.

$sel:endTime:ImportJobProperties', importJobProperties_endTime - The time that the Import job was completed.

$sel:dataAccessRoleArn:ImportJobProperties', importJobProperties_dataAccessRoleArn - The Amazon Resource Name (ARN) that gives Amazon HealthLake access to your input data.

$sel:message:ImportJobProperties', importJobProperties_message - An explanation of any errors that may have occurred during the FHIR import job.

$sel:jobId:ImportJobProperties', importJobProperties_jobId - The AWS-generated id number for the Import job.

$sel:jobStatus:ImportJobProperties', importJobProperties_jobStatus - The job status for an Import job. Possible statuses are SUBMITTED, IN_PROGRESS, COMPLETED, FAILED.

$sel:submitTime:ImportJobProperties', importJobProperties_submitTime - The time that the Import job was submitted for processing.

$sel:datastoreId:ImportJobProperties', importJobProperties_datastoreId - The datastore id used when the Import job was created.

$sel:inputDataConfig:ImportJobProperties', importJobProperties_inputDataConfig - The input data configuration that was supplied when the Import job was created.

InputDataConfig

data InputDataConfig

The input properties for an import job.

See: newInputDataConfig smart constructor.

Constructors

InputDataConfig' (Maybe Text)

newInputDataConfig :: InputDataConfig

Create a value of InputDataConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Uri:InputDataConfig', inputDataConfig_s3Uri - The S3Uri is the user specified S3 location of the FHIR data to be imported into Amazon HealthLake.

KmsEncryptionConfig

data KmsEncryptionConfig

The customer-managed-key(CMK) used when creating a Data Store. If a customer owned key is not specified, an AWS owned key will be used for encryption.

See: newKmsEncryptionConfig smart constructor.

Constructors

KmsEncryptionConfig' (Maybe Text) CmkType

newKmsEncryptionConfig

Arguments

:: CmkType
$sel:cmkType:KmsEncryptionConfig'

-> KmsEncryptionConfig

Create a value of KmsEncryptionConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyId:KmsEncryptionConfig', kmsEncryptionConfig_kmsKeyId - The KMS encryption key id/alias used to encrypt the Data Store contents at rest.

$sel:cmkType:KmsEncryptionConfig', kmsEncryptionConfig_cmkType - The type of customer-managed-key(CMK) used for encyrption. The two types of supported CMKs are customer owned CMKs and AWS owned CMKs.

OutputDataConfig

data OutputDataConfig

The output data configuration that was supplied when the export job was created.

See: newOutputDataConfig smart constructor.

Constructors

OutputDataConfig' (Maybe S3Configuration)

newOutputDataConfig :: OutputDataConfig

Create a value of OutputDataConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Configuration:OutputDataConfig', outputDataConfig_s3Configuration - The output data configuration that was supplied when the export job was created.

PreloadDataConfig

data PreloadDataConfig

The input properties for the preloaded Data Store. Only data preloaded from Synthea is supported.

See: newPreloadDataConfig smart constructor.

Constructors

PreloadDataConfig' PreloadDataType

newPreloadDataConfig

Arguments

:: PreloadDataType
$sel:preloadDataType:PreloadDataConfig'

-> PreloadDataConfig

Create a value of PreloadDataConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:preloadDataType:PreloadDataConfig', preloadDataConfig_preloadDataType - The type of preloaded data. Only Synthea preloaded data is supported.

S3Configuration

data S3Configuration

The configuration of the S3 bucket for either an import or export job. This includes assigning permissions for access.

See: newS3Configuration smart constructor.

Constructors

S3Configuration' Text Text

newS3Configuration

Arguments

:: Text
$sel:s3Uri:S3Configuration'

-> Text
$sel:kmsKeyId:S3Configuration'

-> S3Configuration

Create a value of S3Configuration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Uri:S3Configuration', s3Configuration_s3Uri - The S3Uri is the user specified S3 location of the FHIR data to be imported into Amazon HealthLake.

$sel:kmsKeyId:S3Configuration', s3Configuration_kmsKeyId - The KMS key ID used to access the S3 bucket.

SseConfiguration

data SseConfiguration

The server-side encryption key configuration for a customer provided encryption key.

See: newSseConfiguration smart constructor.

Constructors

SseConfiguration' KmsEncryptionConfig

newSseConfiguration

Arguments

:: KmsEncryptionConfig
$sel:kmsEncryptionConfig:SseConfiguration'

-> SseConfiguration

Create a value of SseConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsEncryptionConfig:SseConfiguration', sseConfiguration_kmsEncryptionConfig - The KMS encryption configuration used to provide details for data encryption.

Tag

data Tag

A tag is a label consisting of a user-defined key and value. The form for tags is {"Key", "Value"}

See: newTag smart constructor.

Constructors

Tag' Text Text

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The key portion of a tag. Tag keys are case sensitive.

$sel:value:Tag', tag_value - The value portion of tag. Tag values are case sensitive.