Network.AWS.HealthLake.DeleteFHIRDatastore

Description

Deletes a Data Store.

Creating a Request

data DeleteFHIRDatastore

See: newDeleteFHIRDatastore smart constructor.

Constructors

DeleteFHIRDatastore'

Fields

datastoreId :: Maybe Text

The AWS-generated ID for the Data Store to be deleted.

newDeleteFHIRDatastore :: DeleteFHIRDatastore

Create a value of DeleteFHIRDatastore with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datastoreId:DeleteFHIRDatastore', deleteFHIRDatastore_datastoreId - The AWS-generated ID for the Data Store to be deleted.

Request Lenses

deleteFHIRDatastore_datastoreId :: Lens' DeleteFHIRDatastore (Maybe Text)

The AWS-generated ID for the Data Store to be deleted.

Destructuring the Response

data DeleteFHIRDatastoreResponse

See: newDeleteFHIRDatastoreResponse smart constructor.

Constructors

DeleteFHIRDatastoreResponse'

Fields

httpStatus :: Int

The response's http status code.

datastoreId :: Text

The AWS-generated ID for the Data Store to be deleted.

datastoreArn :: Text

The Amazon Resource Name (ARN) that gives Amazon HealthLake access permission.

datastoreStatus :: DatastoreStatus

The status of the Data Store that the user has requested to be deleted.

datastoreEndpoint :: Text

The AWS endpoint for the Data Store the user has requested to be deleted.

newDeleteFHIRDatastoreResponse

Arguments

:: Int
$sel:httpStatus:DeleteFHIRDatastoreResponse'

-> Text
$sel:datastoreId:DeleteFHIRDatastore'

-> Text
$sel:datastoreArn:DeleteFHIRDatastoreResponse'

-> DatastoreStatus
$sel:datastoreStatus:DeleteFHIRDatastoreResponse'

-> Text
$sel:datastoreEndpoint:DeleteFHIRDatastoreResponse'

-> DeleteFHIRDatastoreResponse

Create a value of DeleteFHIRDatastoreResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteFHIRDatastoreResponse', deleteFHIRDatastoreResponse_httpStatus - The response's http status code.

$sel:datastoreId:DeleteFHIRDatastore', deleteFHIRDatastoreResponse_datastoreId - The AWS-generated ID for the Data Store to be deleted.

$sel:datastoreArn:DeleteFHIRDatastoreResponse', deleteFHIRDatastoreResponse_datastoreArn - The Amazon Resource Name (ARN) that gives Amazon HealthLake access permission.

$sel:datastoreStatus:DeleteFHIRDatastoreResponse', deleteFHIRDatastoreResponse_datastoreStatus - The status of the Data Store that the user has requested to be deleted.

$sel:datastoreEndpoint:DeleteFHIRDatastoreResponse', deleteFHIRDatastoreResponse_datastoreEndpoint - The AWS endpoint for the Data Store the user has requested to be deleted.

Response Lenses

deleteFHIRDatastoreResponse_httpStatus :: Lens' DeleteFHIRDatastoreResponse Int

The response's http status code.

deleteFHIRDatastoreResponse_datastoreId :: Lens' DeleteFHIRDatastoreResponse Text

The AWS-generated ID for the Data Store to be deleted.

deleteFHIRDatastoreResponse_datastoreArn :: Lens' DeleteFHIRDatastoreResponse Text

The Amazon Resource Name (ARN) that gives Amazon HealthLake access permission.

deleteFHIRDatastoreResponse_datastoreStatus :: Lens' DeleteFHIRDatastoreResponse DatastoreStatus

The status of the Data Store that the user has requested to be deleted.

deleteFHIRDatastoreResponse_datastoreEndpoint :: Lens' DeleteFHIRDatastoreResponse Text

The AWS endpoint for the Data Store the user has requested to be deleted.