Network.AWS.HealthLake.Lens

Description

Operations

StartFHIRImportJob

startFHIRImportJob_jobName :: Lens' StartFHIRImportJob (Maybe Text)

The name of the FHIR Import job in the StartFHIRImport job request.

startFHIRImportJob_inputDataConfig :: Lens' StartFHIRImportJob InputDataConfig

The input properties of the FHIR Import job in the StartFHIRImport job request.

startFHIRImportJob_jobOutputDataConfig :: Lens' StartFHIRImportJob OutputDataConfig

Undocumented member.

startFHIRImportJob_datastoreId :: Lens' StartFHIRImportJob Text

The AWS-generated Data Store ID.

startFHIRImportJob_dataAccessRoleArn :: Lens' StartFHIRImportJob Text

The Amazon Resource Name (ARN) that gives Amazon HealthLake access permission.

startFHIRImportJob_clientToken :: Lens' StartFHIRImportJob Text

Optional user provided token used for ensuring idempotency.

startFHIRImportJobResponse_datastoreId :: Lens' StartFHIRImportJobResponse (Maybe Text)

The AWS-generated Data Store ID.

startFHIRImportJobResponse_httpStatus :: Lens' StartFHIRImportJobResponse Int

The response's http status code.

startFHIRImportJobResponse_jobId :: Lens' StartFHIRImportJobResponse Text

The AWS-generated job ID.

startFHIRImportJobResponse_jobStatus :: Lens' StartFHIRImportJobResponse JobStatus

The status of an import job.

DescribeFHIRDatastore

describeFHIRDatastore_datastoreId :: Lens' DescribeFHIRDatastore (Maybe Text)

The AWS-generated Data Store id. This is part of the ‘CreateFHIRDatastore’ output.

describeFHIRDatastoreResponse_httpStatus :: Lens' DescribeFHIRDatastoreResponse Int

The response's http status code.

describeFHIRDatastoreResponse_datastoreProperties :: Lens' DescribeFHIRDatastoreResponse DatastoreProperties

All properties associated with a Data Store, including the Data Store ID, Data Store ARN, Data Store name, Data Store status, created at, Data Store type version, and Data Store endpoint.

DescribeFHIRImportJob

describeFHIRImportJob_datastoreId :: Lens' DescribeFHIRImportJob Text

The AWS-generated ID of the Data Store.

describeFHIRImportJob_jobId :: Lens' DescribeFHIRImportJob Text

The AWS-generated job ID.

describeFHIRImportJobResponse_httpStatus :: Lens' DescribeFHIRImportJobResponse Int

The response's http status code.

describeFHIRImportJobResponse_importJobProperties :: Lens' DescribeFHIRImportJobResponse ImportJobProperties

The properties of the Import job request, including the ID, ARN, name, and the status of the job.

ListTagsForResource

listTagsForResource_resourceARN :: Lens' ListTagsForResource Text

The Amazon Resource Name(ARN) of the Data Store for which tags are being added.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe [Tag])

Returns a list of tags associated with a Data Store.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

DeleteFHIRDatastore

deleteFHIRDatastore_datastoreId :: Lens' DeleteFHIRDatastore (Maybe Text)

The AWS-generated ID for the Data Store to be deleted.

deleteFHIRDatastoreResponse_httpStatus :: Lens' DeleteFHIRDatastoreResponse Int

The response's http status code.

deleteFHIRDatastoreResponse_datastoreId :: Lens' DeleteFHIRDatastoreResponse Text

The AWS-generated ID for the Data Store to be deleted.

deleteFHIRDatastoreResponse_datastoreArn :: Lens' DeleteFHIRDatastoreResponse Text

The Amazon Resource Name (ARN) that gives Amazon HealthLake access permission.

deleteFHIRDatastoreResponse_datastoreStatus :: Lens' DeleteFHIRDatastoreResponse DatastoreStatus

The status of the Data Store that the user has requested to be deleted.

deleteFHIRDatastoreResponse_datastoreEndpoint :: Lens' DeleteFHIRDatastoreResponse Text

The AWS endpoint for the Data Store the user has requested to be deleted.

DescribeFHIRExportJob

describeFHIRExportJob_datastoreId :: Lens' DescribeFHIRExportJob Text

The AWS generated ID for the Data Store from which files are being exported from for an export job.

describeFHIRExportJob_jobId :: Lens' DescribeFHIRExportJob Text

The AWS generated ID for an export job.

describeFHIRExportJobResponse_httpStatus :: Lens' DescribeFHIRExportJobResponse Int

The response's http status code.

describeFHIRExportJobResponse_exportJobProperties :: Lens' DescribeFHIRExportJobResponse ExportJobProperties

Displays the properties of the export job, including the ID, Arn, Name, and the status of the job.

CreateFHIRDatastore

createFHIRDatastore_clientToken :: Lens' CreateFHIRDatastore (Maybe Text)

Optional user provided token used for ensuring idempotency.

createFHIRDatastore_sseConfiguration :: Lens' CreateFHIRDatastore (Maybe SseConfiguration)

The server-side encryption key configuration for a customer provided encryption key specified for creating a Data Store.

createFHIRDatastore_datastoreName :: Lens' CreateFHIRDatastore (Maybe Text)

The user generated name for the Data Store.

createFHIRDatastore_preloadDataConfig :: Lens' CreateFHIRDatastore (Maybe PreloadDataConfig)

Optional parameter to preload data upon creation of the Data Store. Currently, the only supported preloaded data is synthetic data generated from Synthea.

createFHIRDatastore_tags :: Lens' CreateFHIRDatastore (Maybe [Tag])

Resource tags that are applied to a Data Store when it is created.

createFHIRDatastore_datastoreTypeVersion :: Lens' CreateFHIRDatastore FHIRVersion

The FHIR version of the Data Store. The only supported version is R4.

createFHIRDatastoreResponse_httpStatus :: Lens' CreateFHIRDatastoreResponse Int

The response's http status code.

createFHIRDatastoreResponse_datastoreId :: Lens' CreateFHIRDatastoreResponse Text

The AWS-generated Data Store id. This id is in the output from the initial Data Store creation call.

createFHIRDatastoreResponse_datastoreArn :: Lens' CreateFHIRDatastoreResponse Text

The datastore ARN is generated during the creation of the Data Store and can be found in the output from the initial Data Store creation call.

createFHIRDatastoreResponse_datastoreStatus :: Lens' CreateFHIRDatastoreResponse DatastoreStatus

The status of the FHIR Data Store. Possible statuses are ‘CREATING’, ‘ACTIVE’, ‘DELETING’, ‘DELETED’.

createFHIRDatastoreResponse_datastoreEndpoint :: Lens' CreateFHIRDatastoreResponse Text

The AWS endpoint for the created Data Store. For preview, only US-east-1 endpoints are supported.

ListFHIRExportJobs

listFHIRExportJobs_submittedAfter :: Lens' ListFHIRExportJobs (Maybe UTCTime)

This parameter limits the response to FHIR export jobs submitted after a user specified date.

listFHIRExportJobs_jobName :: Lens' ListFHIRExportJobs (Maybe Text)

This parameter limits the response to the export job with the specified job name.

listFHIRExportJobs_submittedBefore :: Lens' ListFHIRExportJobs (Maybe UTCTime)

This parameter limits the response to FHIR export jobs submitted before a user specified date.

listFHIRExportJobs_nextToken :: Lens' ListFHIRExportJobs (Maybe Text)

A pagination token used to identify the next page of results to return for a ListFHIRExportJobs query.

listFHIRExportJobs_jobStatus :: Lens' ListFHIRExportJobs (Maybe JobStatus)

This parameter limits the response to the export jobs with the specified job status.

listFHIRExportJobs_maxResults :: Lens' ListFHIRExportJobs (Maybe Natural)

This parameter limits the number of results returned for a ListFHIRExportJobs to a maximum quantity specified by the user.

listFHIRExportJobs_datastoreId :: Lens' ListFHIRExportJobs Text

This parameter limits the response to the export job with the specified Data Store ID.

listFHIRExportJobsResponse_nextToken :: Lens' ListFHIRExportJobsResponse (Maybe Text)

A pagination token used to identify the next page of results to return for a ListFHIRExportJobs query.

listFHIRExportJobsResponse_httpStatus :: Lens' ListFHIRExportJobsResponse Int

The response's http status code.

listFHIRExportJobsResponse_exportJobPropertiesList :: Lens' ListFHIRExportJobsResponse [ExportJobProperties]

The properties of listed FHIR export jobs, including the ID, ARN, name, and the status of the job.

TagResource

tagResource_resourceARN :: Lens' TagResource Text

The Amazon Resource Name(ARN)that gives Amazon HealthLake access to the Data Store which tags are being added to.

tagResource_tags :: Lens' TagResource [Tag]

The user specified key and value pair tags being added to a Data Store.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_resourceARN :: Lens' UntagResource Text

"The Amazon Resource Name(ARN) of the Data Store for which tags are being removed

untagResource_tagKeys :: Lens' UntagResource [Text]

The keys for the tags to be removed from the Healthlake Data Store.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

ListFHIRDatastores

listFHIRDatastores_nextToken :: Lens' ListFHIRDatastores (Maybe Text)

Fetches the next page of Data Stores when results are paginated.

listFHIRDatastores_filter :: Lens' ListFHIRDatastores (Maybe DatastoreFilter)

Lists all filters associated with a FHIR Data Store request.

listFHIRDatastores_maxResults :: Lens' ListFHIRDatastores (Maybe Natural)

The maximum number of Data Stores returned in a single page of a ListFHIRDatastoresRequest call.

listFHIRDatastoresResponse_nextToken :: Lens' ListFHIRDatastoresResponse (Maybe Text)

Pagination token that can be used to retrieve the next page of results.

listFHIRDatastoresResponse_httpStatus :: Lens' ListFHIRDatastoresResponse Int

The response's http status code.

listFHIRDatastoresResponse_datastorePropertiesList :: Lens' ListFHIRDatastoresResponse [DatastoreProperties]

All properties associated with the listed Data Stores.

StartFHIRExportJob

startFHIRExportJob_jobName :: Lens' StartFHIRExportJob (Maybe Text)

The user generated name for an export job.

startFHIRExportJob_outputDataConfig :: Lens' StartFHIRExportJob OutputDataConfig

The output data configuration that was supplied when the export job was created.

startFHIRExportJob_datastoreId :: Lens' StartFHIRExportJob Text

The AWS generated ID for the Data Store from which files are being exported for an export job.

startFHIRExportJob_dataAccessRoleArn :: Lens' StartFHIRExportJob Text

The Amazon Resource Name used during the initiation of the job.

startFHIRExportJob_clientToken :: Lens' StartFHIRExportJob Text

An optional user provided token used for ensuring idempotency.

startFHIRExportJobResponse_datastoreId :: Lens' StartFHIRExportJobResponse (Maybe Text)

The AWS generated ID for the Data Store from which files are being exported for an export job.

startFHIRExportJobResponse_httpStatus :: Lens' StartFHIRExportJobResponse Int

The response's http status code.

startFHIRExportJobResponse_jobId :: Lens' StartFHIRExportJobResponse Text

The AWS generated ID for an export job.

startFHIRExportJobResponse_jobStatus :: Lens' StartFHIRExportJobResponse JobStatus

The status of a FHIR export job. Possible statuses are SUBMITTED, IN_PROGRESS, COMPLETED, or FAILED.

ListFHIRImportJobs

listFHIRImportJobs_submittedAfter :: Lens' ListFHIRImportJobs (Maybe UTCTime)

This parameter limits the response to FHIR import jobs submitted after a user specified date.

listFHIRImportJobs_jobName :: Lens' ListFHIRImportJobs (Maybe Text)

This parameter limits the response to the import job with the specified job name.

listFHIRImportJobs_submittedBefore :: Lens' ListFHIRImportJobs (Maybe UTCTime)

This parameter limits the response to FHIR import jobs submitted before a user specified date.

listFHIRImportJobs_nextToken :: Lens' ListFHIRImportJobs (Maybe Text)

A pagination token used to identify the next page of results to return for a ListFHIRImportJobs query.

listFHIRImportJobs_jobStatus :: Lens' ListFHIRImportJobs (Maybe JobStatus)

This parameter limits the response to the import job with the specified job status.

listFHIRImportJobs_maxResults :: Lens' ListFHIRImportJobs (Maybe Natural)

This parameter limits the number of results returned for a ListFHIRImportJobs to a maximum quantity specified by the user.

listFHIRImportJobs_datastoreId :: Lens' ListFHIRImportJobs Text

This parameter limits the response to the import job with the specified Data Store ID.

listFHIRImportJobsResponse_nextToken :: Lens' ListFHIRImportJobsResponse (Maybe Text)

A pagination token used to identify the next page of results to return for a ListFHIRImportJobs query.

listFHIRImportJobsResponse_httpStatus :: Lens' ListFHIRImportJobsResponse Int

The response's http status code.

listFHIRImportJobsResponse_importJobPropertiesList :: Lens' ListFHIRImportJobsResponse [ImportJobProperties]

The properties of a listed FHIR import jobs, including the ID, ARN, name, and the status of the job.

Types

DatastoreFilter

datastoreFilter_createdAfter :: Lens' DatastoreFilter (Maybe UTCTime)

A filter that allows the user to set cutoff dates for records. All Data Stores created after the specified date will be included in the results.

datastoreFilter_datastoreName :: Lens' DatastoreFilter (Maybe Text)

Allows the user to filter Data Store results by name.

datastoreFilter_datastoreStatus :: Lens' DatastoreFilter (Maybe DatastoreStatus)

Allows the user to filter Data Store results by status.

datastoreFilter_createdBefore :: Lens' DatastoreFilter (Maybe UTCTime)

A filter that allows the user to set cutoff dates for records. All Data Stores created before the specified date will be included in the results.

DatastoreProperties

datastoreProperties_sseConfiguration :: Lens' DatastoreProperties (Maybe SseConfiguration)

The server-side encryption key configuration for a customer provided encryption key (CMK).

datastoreProperties_createdAt :: Lens' DatastoreProperties (Maybe UTCTime)

The time that a Data Store was created.

datastoreProperties_datastoreName :: Lens' DatastoreProperties (Maybe Text)

The user-generated name for the Data Store.

datastoreProperties_preloadDataConfig :: Lens' DatastoreProperties (Maybe PreloadDataConfig)

The preloaded data configuration for the Data Store. Only data preloaded from Synthea is supported.

datastoreProperties_datastoreId :: Lens' DatastoreProperties Text

The AWS-generated ID number for the Data Store.

datastoreProperties_datastoreArn :: Lens' DatastoreProperties Text

The Amazon Resource Name used in the creation of the Data Store.

datastoreProperties_datastoreStatus :: Lens' DatastoreProperties DatastoreStatus

The status of the Data Store. Possible statuses are 'CREATING', 'ACTIVE', 'DELETING', or 'DELETED'.

datastoreProperties_datastoreTypeVersion :: Lens' DatastoreProperties FHIRVersion

The FHIR version. Only R4 version data is supported.

datastoreProperties_datastoreEndpoint :: Lens' DatastoreProperties Text

The AWS endpoint for the Data Store. Each Data Store will have it's own endpoint with Data Store ID in the endpoint URL.

ExportJobProperties

exportJobProperties_jobName :: Lens' ExportJobProperties (Maybe Text)

The user generated name for an export job.

exportJobProperties_endTime :: Lens' ExportJobProperties (Maybe UTCTime)

The time an export job completed.

exportJobProperties_dataAccessRoleArn :: Lens' ExportJobProperties (Maybe Text)

The Amazon Resource Name used during the initiation of the job.

exportJobProperties_message :: Lens' ExportJobProperties (Maybe Text)

An explanation of any errors that may have occurred during the export job.

exportJobProperties_jobId :: Lens' ExportJobProperties Text

The AWS generated ID for an export job.

exportJobProperties_jobStatus :: Lens' ExportJobProperties JobStatus

The status of a FHIR export job. Possible statuses are SUBMITTED, IN_PROGRESS, COMPLETED, or FAILED.

exportJobProperties_submitTime :: Lens' ExportJobProperties UTCTime

The time an export job was initiated.

exportJobProperties_datastoreId :: Lens' ExportJobProperties Text

The AWS generated ID for the Data Store from which files are being exported for an export job.

exportJobProperties_outputDataConfig :: Lens' ExportJobProperties OutputDataConfig

The output data configuration that was supplied when the export job was created.

ImportJobProperties

importJobProperties_jobOutputDataConfig :: Lens' ImportJobProperties (Maybe OutputDataConfig)

Undocumented member.

importJobProperties_jobName :: Lens' ImportJobProperties (Maybe Text)

The user-generated name for an Import job.

importJobProperties_endTime :: Lens' ImportJobProperties (Maybe UTCTime)

The time that the Import job was completed.

importJobProperties_dataAccessRoleArn :: Lens' ImportJobProperties (Maybe Text)

The Amazon Resource Name (ARN) that gives Amazon HealthLake access to your input data.

importJobProperties_message :: Lens' ImportJobProperties (Maybe Text)

An explanation of any errors that may have occurred during the FHIR import job.

importJobProperties_jobId :: Lens' ImportJobProperties Text

The AWS-generated id number for the Import job.

importJobProperties_jobStatus :: Lens' ImportJobProperties JobStatus

The job status for an Import job. Possible statuses are SUBMITTED, IN_PROGRESS, COMPLETED, FAILED.

importJobProperties_submitTime :: Lens' ImportJobProperties UTCTime

The time that the Import job was submitted for processing.

importJobProperties_datastoreId :: Lens' ImportJobProperties Text

The datastore id used when the Import job was created.

importJobProperties_inputDataConfig :: Lens' ImportJobProperties InputDataConfig

The input data configuration that was supplied when the Import job was created.

InputDataConfig

inputDataConfig_s3Uri :: Lens' InputDataConfig (Maybe Text)

The S3Uri is the user specified S3 location of the FHIR data to be imported into Amazon HealthLake.

KmsEncryptionConfig

kmsEncryptionConfig_kmsKeyId :: Lens' KmsEncryptionConfig (Maybe Text)

The KMS encryption key id/alias used to encrypt the Data Store contents at rest.

kmsEncryptionConfig_cmkType :: Lens' KmsEncryptionConfig CmkType

The type of customer-managed-key(CMK) used for encyrption. The two types of supported CMKs are customer owned CMKs and AWS owned CMKs.

OutputDataConfig

outputDataConfig_s3Configuration :: Lens' OutputDataConfig (Maybe S3Configuration)

The output data configuration that was supplied when the export job was created.

PreloadDataConfig

preloadDataConfig_preloadDataType :: Lens' PreloadDataConfig PreloadDataType

The type of preloaded data. Only Synthea preloaded data is supported.

S3Configuration

s3Configuration_s3Uri :: Lens' S3Configuration Text

The S3Uri is the user specified S3 location of the FHIR data to be imported into Amazon HealthLake.

s3Configuration_kmsKeyId :: Lens' S3Configuration Text

The KMS key ID used to access the S3 bucket.

SseConfiguration

sseConfiguration_kmsEncryptionConfig :: Lens' SseConfiguration KmsEncryptionConfig

The KMS encryption configuration used to provide details for data encryption.

Tag

tag_key :: Lens' Tag Text

The key portion of a tag. Tag keys are case sensitive.

tag_value :: Lens' Tag Text

The value portion of tag. Tag values are case sensitive.