Network.AWS.HealthLake.ListFHIRDatastores

Description

Lists all FHIR Data Stores that are in the user’s account, regardless of Data Store status.

Creating a Request

data ListFHIRDatastores

See: newListFHIRDatastores smart constructor.

Constructors

ListFHIRDatastores'

Fields

nextToken :: Maybe Text

Fetches the next page of Data Stores when results are paginated.

filter' :: Maybe DatastoreFilter

Lists all filters associated with a FHIR Data Store request.

maxResults :: Maybe Natural

The maximum number of Data Stores returned in a single page of a ListFHIRDatastoresRequest call.

newListFHIRDatastores :: ListFHIRDatastores

Create a value of ListFHIRDatastores with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFHIRDatastores', listFHIRDatastores_nextToken - Fetches the next page of Data Stores when results are paginated.

$sel:filter':ListFHIRDatastores', listFHIRDatastores_filter - Lists all filters associated with a FHIR Data Store request.

$sel:maxResults:ListFHIRDatastores', listFHIRDatastores_maxResults - The maximum number of Data Stores returned in a single page of a ListFHIRDatastoresRequest call.

Request Lenses

listFHIRDatastores_nextToken :: Lens' ListFHIRDatastores (Maybe Text)

Fetches the next page of Data Stores when results are paginated.

listFHIRDatastores_filter :: Lens' ListFHIRDatastores (Maybe DatastoreFilter)

Lists all filters associated with a FHIR Data Store request.

listFHIRDatastores_maxResults :: Lens' ListFHIRDatastores (Maybe Natural)

The maximum number of Data Stores returned in a single page of a ListFHIRDatastoresRequest call.

Destructuring the Response

data ListFHIRDatastoresResponse

See: newListFHIRDatastoresResponse smart constructor.

Constructors

ListFHIRDatastoresResponse'

Fields

nextToken :: Maybe Text

Pagination token that can be used to retrieve the next page of results.

httpStatus :: Int

The response's http status code.

datastorePropertiesList :: [DatastoreProperties]

All properties associated with the listed Data Stores.

newListFHIRDatastoresResponse

Arguments

:: Int
$sel:httpStatus:ListFHIRDatastoresResponse'

-> ListFHIRDatastoresResponse

Create a value of ListFHIRDatastoresResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFHIRDatastores', listFHIRDatastoresResponse_nextToken - Pagination token that can be used to retrieve the next page of results.

$sel:httpStatus:ListFHIRDatastoresResponse', listFHIRDatastoresResponse_httpStatus - The response's http status code.

$sel:datastorePropertiesList:ListFHIRDatastoresResponse', listFHIRDatastoresResponse_datastorePropertiesList - All properties associated with the listed Data Stores.

Response Lenses

listFHIRDatastoresResponse_nextToken :: Lens' ListFHIRDatastoresResponse (Maybe Text)

Pagination token that can be used to retrieve the next page of results.

listFHIRDatastoresResponse_httpStatus :: Lens' ListFHIRDatastoresResponse Int

The response's http status code.

listFHIRDatastoresResponse_datastorePropertiesList :: Lens' ListFHIRDatastoresResponse [DatastoreProperties]

All properties associated with the listed Data Stores.