Network.AWS.HealthLake.Types.DatastoreFilter

Description

Documentation

data DatastoreFilter

The filters applied to Data Store query.

See: newDatastoreFilter smart constructor.

Constructors

DatastoreFilter'

Fields

createdAfter :: Maybe POSIX

A filter that allows the user to set cutoff dates for records. All Data Stores created after the specified date will be included in the results.

datastoreName :: Maybe Text

Allows the user to filter Data Store results by name.

datastoreStatus :: Maybe DatastoreStatus

Allows the user to filter Data Store results by status.

createdBefore :: Maybe POSIX

A filter that allows the user to set cutoff dates for records. All Data Stores created before the specified date will be included in the results.

newDatastoreFilter :: DatastoreFilter

Create a value of DatastoreFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAfter:DatastoreFilter', datastoreFilter_createdAfter - A filter that allows the user to set cutoff dates for records. All Data Stores created after the specified date will be included in the results.

$sel:datastoreName:DatastoreFilter', datastoreFilter_datastoreName - Allows the user to filter Data Store results by name.

$sel:datastoreStatus:DatastoreFilter', datastoreFilter_datastoreStatus - Allows the user to filter Data Store results by status.

$sel:createdBefore:DatastoreFilter', datastoreFilter_createdBefore - A filter that allows the user to set cutoff dates for records. All Data Stores created before the specified date will be included in the results.

datastoreFilter_createdAfter :: Lens' DatastoreFilter (Maybe UTCTime)

A filter that allows the user to set cutoff dates for records. All Data Stores created after the specified date will be included in the results.

datastoreFilter_datastoreName :: Lens' DatastoreFilter (Maybe Text)

Allows the user to filter Data Store results by name.

datastoreFilter_datastoreStatus :: Lens' DatastoreFilter (Maybe DatastoreStatus)

Allows the user to filter Data Store results by status.

datastoreFilter_createdBefore :: Lens' DatastoreFilter (Maybe UTCTime)

A filter that allows the user to set cutoff dates for records. All Data Stores created before the specified date will be included in the results.