Network.AWS.HealthLake.Types.ExportJobProperties

Description

Documentation

data ExportJobProperties

The properties of a FHIR export job, including the ID, ARN, name, and the status of the job.

See: newExportJobProperties smart constructor.

Constructors

ExportJobProperties'

Fields

jobName :: Maybe Text

The user generated name for an export job.

endTime :: Maybe POSIX

The time an export job completed.

dataAccessRoleArn :: Maybe Text

The Amazon Resource Name used during the initiation of the job.

message :: Maybe Text

An explanation of any errors that may have occurred during the export job.

jobId :: Text

The AWS generated ID for an export job.

jobStatus :: JobStatus

The status of a FHIR export job. Possible statuses are SUBMITTED, IN_PROGRESS, COMPLETED, or FAILED.

submitTime :: POSIX

The time an export job was initiated.

datastoreId :: Text

The AWS generated ID for the Data Store from which files are being exported for an export job.

outputDataConfig :: OutputDataConfig

The output data configuration that was supplied when the export job was created.

newExportJobProperties

Arguments

:: Text
$sel:jobId:ExportJobProperties'

-> JobStatus
$sel:jobStatus:ExportJobProperties'

-> UTCTime
$sel:submitTime:ExportJobProperties'

-> Text
$sel:datastoreId:ExportJobProperties'

-> OutputDataConfig
$sel:outputDataConfig:ExportJobProperties'

-> ExportJobProperties

Create a value of ExportJobProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobName:ExportJobProperties', exportJobProperties_jobName - The user generated name for an export job.

$sel:endTime:ExportJobProperties', exportJobProperties_endTime - The time an export job completed.

$sel:dataAccessRoleArn:ExportJobProperties', exportJobProperties_dataAccessRoleArn - The Amazon Resource Name used during the initiation of the job.

$sel:message:ExportJobProperties', exportJobProperties_message - An explanation of any errors that may have occurred during the export job.

$sel:jobId:ExportJobProperties', exportJobProperties_jobId - The AWS generated ID for an export job.

$sel:jobStatus:ExportJobProperties', exportJobProperties_jobStatus - The status of a FHIR export job. Possible statuses are SUBMITTED, IN_PROGRESS, COMPLETED, or FAILED.

$sel:submitTime:ExportJobProperties', exportJobProperties_submitTime - The time an export job was initiated.

$sel:datastoreId:ExportJobProperties', exportJobProperties_datastoreId - The AWS generated ID for the Data Store from which files are being exported for an export job.

$sel:outputDataConfig:ExportJobProperties', exportJobProperties_outputDataConfig - The output data configuration that was supplied when the export job was created.

exportJobProperties_jobName :: Lens' ExportJobProperties (Maybe Text)

The user generated name for an export job.

exportJobProperties_endTime :: Lens' ExportJobProperties (Maybe UTCTime)

The time an export job completed.

exportJobProperties_dataAccessRoleArn :: Lens' ExportJobProperties (Maybe Text)

The Amazon Resource Name used during the initiation of the job.

exportJobProperties_message :: Lens' ExportJobProperties (Maybe Text)

An explanation of any errors that may have occurred during the export job.

exportJobProperties_jobId :: Lens' ExportJobProperties Text

The AWS generated ID for an export job.

exportJobProperties_jobStatus :: Lens' ExportJobProperties JobStatus

The status of a FHIR export job. Possible statuses are SUBMITTED, IN_PROGRESS, COMPLETED, or FAILED.

exportJobProperties_submitTime :: Lens' ExportJobProperties UTCTime

The time an export job was initiated.

exportJobProperties_datastoreId :: Lens' ExportJobProperties Text

The AWS generated ID for the Data Store from which files are being exported for an export job.

exportJobProperties_outputDataConfig :: Lens' ExportJobProperties OutputDataConfig

The output data configuration that was supplied when the export job was created.