Network.AWS.HealthLake.Types.ImportJobProperties

Description

Documentation

data ImportJobProperties

Displays the properties of the import job, including the ID, Arn, Name, and the status of the Data Store.

See: newImportJobProperties smart constructor.

Constructors

ImportJobProperties'

Fields

jobOutputDataConfig :: Maybe OutputDataConfig

jobName :: Maybe Text

The user-generated name for an Import job.

endTime :: Maybe POSIX

The time that the Import job was completed.

dataAccessRoleArn :: Maybe Text

The Amazon Resource Name (ARN) that gives Amazon HealthLake access to your input data.

message :: Maybe Text

An explanation of any errors that may have occurred during the FHIR import job.

jobId :: Text

The AWS-generated id number for the Import job.

jobStatus :: JobStatus

The job status for an Import job. Possible statuses are SUBMITTED, IN_PROGRESS, COMPLETED, FAILED.

submitTime :: POSIX

The time that the Import job was submitted for processing.

datastoreId :: Text

The datastore id used when the Import job was created.

inputDataConfig :: InputDataConfig

The input data configuration that was supplied when the Import job was created.

newImportJobProperties

Arguments

:: Text
$sel:jobId:ImportJobProperties'

-> JobStatus
$sel:jobStatus:ImportJobProperties'

-> UTCTime
$sel:submitTime:ImportJobProperties'

-> Text
$sel:datastoreId:ImportJobProperties'

-> InputDataConfig
$sel:inputDataConfig:ImportJobProperties'

-> ImportJobProperties

Create a value of ImportJobProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobOutputDataConfig:ImportJobProperties', importJobProperties_jobOutputDataConfig - Undocumented member.

$sel:jobName:ImportJobProperties', importJobProperties_jobName - The user-generated name for an Import job.

$sel:endTime:ImportJobProperties', importJobProperties_endTime - The time that the Import job was completed.

$sel:dataAccessRoleArn:ImportJobProperties', importJobProperties_dataAccessRoleArn - The Amazon Resource Name (ARN) that gives Amazon HealthLake access to your input data.

$sel:message:ImportJobProperties', importJobProperties_message - An explanation of any errors that may have occurred during the FHIR import job.

$sel:jobId:ImportJobProperties', importJobProperties_jobId - The AWS-generated id number for the Import job.

$sel:jobStatus:ImportJobProperties', importJobProperties_jobStatus - The job status for an Import job. Possible statuses are SUBMITTED, IN_PROGRESS, COMPLETED, FAILED.

$sel:submitTime:ImportJobProperties', importJobProperties_submitTime - The time that the Import job was submitted for processing.

$sel:datastoreId:ImportJobProperties', importJobProperties_datastoreId - The datastore id used when the Import job was created.

$sel:inputDataConfig:ImportJobProperties', importJobProperties_inputDataConfig - The input data configuration that was supplied when the Import job was created.

importJobProperties_jobOutputDataConfig :: Lens' ImportJobProperties (Maybe OutputDataConfig)

Undocumented member.

importJobProperties_jobName :: Lens' ImportJobProperties (Maybe Text)

The user-generated name for an Import job.

importJobProperties_endTime :: Lens' ImportJobProperties (Maybe UTCTime)

The time that the Import job was completed.

importJobProperties_dataAccessRoleArn :: Lens' ImportJobProperties (Maybe Text)

The Amazon Resource Name (ARN) that gives Amazon HealthLake access to your input data.

importJobProperties_message :: Lens' ImportJobProperties (Maybe Text)

An explanation of any errors that may have occurred during the FHIR import job.

importJobProperties_jobId :: Lens' ImportJobProperties Text

The AWS-generated id number for the Import job.

importJobProperties_jobStatus :: Lens' ImportJobProperties JobStatus

The job status for an Import job. Possible statuses are SUBMITTED, IN_PROGRESS, COMPLETED, FAILED.

importJobProperties_submitTime :: Lens' ImportJobProperties UTCTime

The time that the Import job was submitted for processing.

importJobProperties_datastoreId :: Lens' ImportJobProperties Text

The datastore id used when the Import job was created.

importJobProperties_inputDataConfig :: Lens' ImportJobProperties InputDataConfig

The input data configuration that was supplied when the Import job was created.