Network.AWS.HealthLake.Types.SseConfiguration

Description

Documentation

data SseConfiguration

The server-side encryption key configuration for a customer provided encryption key.

See: newSseConfiguration smart constructor.

Constructors

SseConfiguration'

Fields

kmsEncryptionConfig :: KmsEncryptionConfig

The KMS encryption configuration used to provide details for data encryption.

newSseConfiguration

Arguments

:: KmsEncryptionConfig
$sel:kmsEncryptionConfig:SseConfiguration'

-> SseConfiguration

Create a value of SseConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsEncryptionConfig:SseConfiguration', sseConfiguration_kmsEncryptionConfig - The KMS encryption configuration used to provide details for data encryption.

sseConfiguration_kmsEncryptionConfig :: Lens' SseConfiguration KmsEncryptionConfig

The KMS encryption configuration used to provide details for data encryption.