Network.AWS.HoneyCode.BatchCreateTableRows

Description

The BatchCreateTableRows API allows you to create one or more rows at the end of a table in a workbook. The API allows you to specify the values to set in some or all of the columns in the new rows.

If a column is not explicitly set in a specific row, then the column level formula specified in the table will be applied to the new row. If there is no column level formula but the last row of the table has a formula, then that formula will be copied down to the new row. If there is no column level formula and no formula in the last row of the table, then that column will be left blank for the new rows.

Creating a Request

data BatchCreateTableRows

See: newBatchCreateTableRows smart constructor.

Constructors

BatchCreateTableRows'

Fields

clientRequestToken :: Maybe Text

The request token for performing the batch create operation. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will not perform the operation again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

workbookId :: Text

The ID of the workbook where the new rows are being added.

If a workbook with the specified ID could not be found, this API throws ResourceNotFoundException.

tableId :: Text

The ID of the table where the new rows are being added.

If a table with the specified ID could not be found, this API throws ResourceNotFoundException.

rowsToCreate :: NonEmpty CreateRowData

The list of rows to create at the end of the table. Each item in this list needs to have a batch item id to uniquely identify the element in the request and the cells to create for that row. You need to specify at least one item in this list.

Note that if one of the column ids in any of the rows in the request does not exist in the table, then the request fails and no updates are made to the table.

newBatchCreateTableRows

Arguments

:: Text
$sel:workbookId:BatchCreateTableRows'

-> Text
$sel:tableId:BatchCreateTableRows'

-> NonEmpty CreateRowData
$sel:rowsToCreate:BatchCreateTableRows'

-> BatchCreateTableRows

Create a value of BatchCreateTableRows with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientRequestToken:BatchCreateTableRows', batchCreateTableRows_clientRequestToken - The request token for performing the batch create operation. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will not perform the operation again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

$sel:workbookId:BatchCreateTableRows', batchCreateTableRows_workbookId - The ID of the workbook where the new rows are being added.

If a workbook with the specified ID could not be found, this API throws ResourceNotFoundException.

$sel:tableId:BatchCreateTableRows', batchCreateTableRows_tableId - The ID of the table where the new rows are being added.

If a table with the specified ID could not be found, this API throws ResourceNotFoundException.

$sel:rowsToCreate:BatchCreateTableRows', batchCreateTableRows_rowsToCreate - The list of rows to create at the end of the table. Each item in this list needs to have a batch item id to uniquely identify the element in the request and the cells to create for that row. You need to specify at least one item in this list.

Note that if one of the column ids in any of the rows in the request does not exist in the table, then the request fails and no updates are made to the table.

Request Lenses

batchCreateTableRows_clientRequestToken :: Lens' BatchCreateTableRows (Maybe Text)

The request token for performing the batch create operation. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will not perform the operation again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

batchCreateTableRows_workbookId :: Lens' BatchCreateTableRows Text

The ID of the workbook where the new rows are being added.

If a workbook with the specified ID could not be found, this API throws ResourceNotFoundException.

batchCreateTableRows_tableId :: Lens' BatchCreateTableRows Text

The ID of the table where the new rows are being added.

If a table with the specified ID could not be found, this API throws ResourceNotFoundException.

batchCreateTableRows_rowsToCreate :: Lens' BatchCreateTableRows (NonEmpty CreateRowData)

The list of rows to create at the end of the table. Each item in this list needs to have a batch item id to uniquely identify the element in the request and the cells to create for that row. You need to specify at least one item in this list.

Note that if one of the column ids in any of the rows in the request does not exist in the table, then the request fails and no updates are made to the table.

Destructuring the Response

data BatchCreateTableRowsResponse

See: newBatchCreateTableRowsResponse smart constructor.

Constructors

BatchCreateTableRowsResponse'

Fields

failedBatchItems :: Maybe [FailedBatchItem]

The list of batch items in the request that could not be added to the table. Each element in this list contains one item from the request that could not be added to the table along with the reason why that item could not be added.

httpStatus :: Int

The response's http status code.

workbookCursor :: Integer

The updated workbook cursor after adding the new rows at the end of the table.

createdRows :: HashMap Text Text

The map of batch item id to the row id that was created for that item.

newBatchCreateTableRowsResponse

Arguments

:: Int
$sel:httpStatus:BatchCreateTableRowsResponse'

-> Integer
$sel:workbookCursor:BatchCreateTableRowsResponse'

-> BatchCreateTableRowsResponse

Create a value of BatchCreateTableRowsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedBatchItems:BatchCreateTableRowsResponse', batchCreateTableRowsResponse_failedBatchItems - The list of batch items in the request that could not be added to the table. Each element in this list contains one item from the request that could not be added to the table along with the reason why that item could not be added.

$sel:httpStatus:BatchCreateTableRowsResponse', batchCreateTableRowsResponse_httpStatus - The response's http status code.

$sel:workbookCursor:BatchCreateTableRowsResponse', batchCreateTableRowsResponse_workbookCursor - The updated workbook cursor after adding the new rows at the end of the table.

$sel:createdRows:BatchCreateTableRowsResponse', batchCreateTableRowsResponse_createdRows - The map of batch item id to the row id that was created for that item.

Response Lenses

batchCreateTableRowsResponse_failedBatchItems :: Lens' BatchCreateTableRowsResponse (Maybe [FailedBatchItem])

The list of batch items in the request that could not be added to the table. Each element in this list contains one item from the request that could not be added to the table along with the reason why that item could not be added.

batchCreateTableRowsResponse_httpStatus :: Lens' BatchCreateTableRowsResponse Int

The response's http status code.

batchCreateTableRowsResponse_workbookCursor :: Lens' BatchCreateTableRowsResponse Integer

The updated workbook cursor after adding the new rows at the end of the table.

batchCreateTableRowsResponse_createdRows :: Lens' BatchCreateTableRowsResponse (HashMap Text Text)

The map of batch item id to the row id that was created for that item.