Network.AWS.HoneyCode.BatchDeleteTableRows

Description

The BatchDeleteTableRows API allows you to delete one or more rows from a table in a workbook. You need to specify the ids of the rows that you want to delete from the table.

Creating a Request

data BatchDeleteTableRows

See: newBatchDeleteTableRows smart constructor.

Constructors

BatchDeleteTableRows'

Fields

clientRequestToken :: Maybe Text

The request token for performing the delete action. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will not perform the action again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

workbookId :: Text

The ID of the workbook where the rows are being deleted.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

tableId :: Text

The ID of the table where the rows are being deleted.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

rowIds :: NonEmpty Text

The list of row ids to delete from the table. You need to specify at least one row id in this list.

Note that if one of the row ids provided in the request does not exist in the table, then the request fails and no rows are deleted from the table.

newBatchDeleteTableRows

Arguments

:: Text
$sel:workbookId:BatchDeleteTableRows'

-> Text
$sel:tableId:BatchDeleteTableRows'

-> NonEmpty Text
$sel:rowIds:BatchDeleteTableRows'

-> BatchDeleteTableRows

Create a value of BatchDeleteTableRows with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientRequestToken:BatchDeleteTableRows', batchDeleteTableRows_clientRequestToken - The request token for performing the delete action. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will not perform the action again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

$sel:workbookId:BatchDeleteTableRows', batchDeleteTableRows_workbookId - The ID of the workbook where the rows are being deleted.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

$sel:tableId:BatchDeleteTableRows', batchDeleteTableRows_tableId - The ID of the table where the rows are being deleted.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

$sel:rowIds:BatchDeleteTableRows', batchDeleteTableRows_rowIds - The list of row ids to delete from the table. You need to specify at least one row id in this list.

Note that if one of the row ids provided in the request does not exist in the table, then the request fails and no rows are deleted from the table.

Request Lenses

batchDeleteTableRows_clientRequestToken :: Lens' BatchDeleteTableRows (Maybe Text)

The request token for performing the delete action. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will not perform the action again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

batchDeleteTableRows_workbookId :: Lens' BatchDeleteTableRows Text

The ID of the workbook where the rows are being deleted.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

batchDeleteTableRows_tableId :: Lens' BatchDeleteTableRows Text

The ID of the table where the rows are being deleted.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

batchDeleteTableRows_rowIds :: Lens' BatchDeleteTableRows (NonEmpty Text)

The list of row ids to delete from the table. You need to specify at least one row id in this list.

Note that if one of the row ids provided in the request does not exist in the table, then the request fails and no rows are deleted from the table.

Destructuring the Response

data BatchDeleteTableRowsResponse

See: newBatchDeleteTableRowsResponse smart constructor.

Constructors

BatchDeleteTableRowsResponse'

Fields

failedBatchItems :: Maybe [FailedBatchItem]

The list of row ids in the request that could not be deleted from the table. Each element in this list contains one row id from the request that could not be deleted along with the reason why that item could not be deleted.

httpStatus :: Int

The response's http status code.

workbookCursor :: Integer

The updated workbook cursor after deleting the rows from the table.

newBatchDeleteTableRowsResponse

Arguments

:: Int
$sel:httpStatus:BatchDeleteTableRowsResponse'

-> Integer
$sel:workbookCursor:BatchDeleteTableRowsResponse'

-> BatchDeleteTableRowsResponse

Create a value of BatchDeleteTableRowsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedBatchItems:BatchDeleteTableRowsResponse', batchDeleteTableRowsResponse_failedBatchItems - The list of row ids in the request that could not be deleted from the table. Each element in this list contains one row id from the request that could not be deleted along with the reason why that item could not be deleted.

$sel:httpStatus:BatchDeleteTableRowsResponse', batchDeleteTableRowsResponse_httpStatus - The response's http status code.

$sel:workbookCursor:BatchDeleteTableRowsResponse', batchDeleteTableRowsResponse_workbookCursor - The updated workbook cursor after deleting the rows from the table.

Response Lenses

batchDeleteTableRowsResponse_failedBatchItems :: Lens' BatchDeleteTableRowsResponse (Maybe [FailedBatchItem])

The list of row ids in the request that could not be deleted from the table. Each element in this list contains one row id from the request that could not be deleted along with the reason why that item could not be deleted.

batchDeleteTableRowsResponse_httpStatus :: Lens' BatchDeleteTableRowsResponse Int

The response's http status code.

batchDeleteTableRowsResponse_workbookCursor :: Lens' BatchDeleteTableRowsResponse Integer

The updated workbook cursor after deleting the rows from the table.