Network.AWS.HoneyCode.BatchUpdateTableRows

Description

The BatchUpdateTableRows API allows you to update one or more rows in a table in a workbook.

You can specify the values to set in some or all of the columns in the table for the specified rows. If a column is not explicitly specified in a particular row, then that column will not be updated for that row. To clear out the data in a specific cell, you need to set the value as an empty string ("").

Creating a Request

data BatchUpdateTableRows

See: newBatchUpdateTableRows smart constructor.

Constructors

BatchUpdateTableRows'

Fields

clientRequestToken :: Maybe Text

The request token for performing the update action. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will not perform the action again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

workbookId :: Text

The ID of the workbook where the rows are being updated.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

tableId :: Text

The ID of the table where the rows are being updated.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

rowsToUpdate :: NonEmpty UpdateRowData

The list of rows to update in the table. Each item in this list needs to contain the row id to update along with the map of column id to cell values for each column in that row that needs to be updated. You need to specify at least one row in this list, and for each row, you need to specify at least one column to update.

Note that if one of the row or column ids in the request does not exist in the table, then the request fails and no updates are made to the table.

newBatchUpdateTableRows

Arguments

:: Text
$sel:workbookId:BatchUpdateTableRows'

-> Text
$sel:tableId:BatchUpdateTableRows'

-> NonEmpty UpdateRowData
$sel:rowsToUpdate:BatchUpdateTableRows'

-> BatchUpdateTableRows

Create a value of BatchUpdateTableRows with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientRequestToken:BatchUpdateTableRows', batchUpdateTableRows_clientRequestToken - The request token for performing the update action. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will not perform the action again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

$sel:workbookId:BatchUpdateTableRows', batchUpdateTableRows_workbookId - The ID of the workbook where the rows are being updated.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

$sel:tableId:BatchUpdateTableRows', batchUpdateTableRows_tableId - The ID of the table where the rows are being updated.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

$sel:rowsToUpdate:BatchUpdateTableRows', batchUpdateTableRows_rowsToUpdate - The list of rows to update in the table. Each item in this list needs to contain the row id to update along with the map of column id to cell values for each column in that row that needs to be updated. You need to specify at least one row in this list, and for each row, you need to specify at least one column to update.

Note that if one of the row or column ids in the request does not exist in the table, then the request fails and no updates are made to the table.

Request Lenses

batchUpdateTableRows_clientRequestToken :: Lens' BatchUpdateTableRows (Maybe Text)

The request token for performing the update action. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will not perform the action again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

batchUpdateTableRows_workbookId :: Lens' BatchUpdateTableRows Text

The ID of the workbook where the rows are being updated.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

batchUpdateTableRows_tableId :: Lens' BatchUpdateTableRows Text

The ID of the table where the rows are being updated.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

batchUpdateTableRows_rowsToUpdate :: Lens' BatchUpdateTableRows (NonEmpty UpdateRowData)

The list of rows to update in the table. Each item in this list needs to contain the row id to update along with the map of column id to cell values for each column in that row that needs to be updated. You need to specify at least one row in this list, and for each row, you need to specify at least one column to update.

Note that if one of the row or column ids in the request does not exist in the table, then the request fails and no updates are made to the table.

Destructuring the Response

data BatchUpdateTableRowsResponse

See: newBatchUpdateTableRowsResponse smart constructor.

Constructors

BatchUpdateTableRowsResponse'

Fields

failedBatchItems :: Maybe [FailedBatchItem]

The list of batch items in the request that could not be updated in the table. Each element in this list contains one item from the request that could not be updated in the table along with the reason why that item could not be updated.

httpStatus :: Int

The response's http status code.

workbookCursor :: Integer

The updated workbook cursor after adding the new rows at the end of the table.

newBatchUpdateTableRowsResponse

Arguments

:: Int
$sel:httpStatus:BatchUpdateTableRowsResponse'

-> Integer
$sel:workbookCursor:BatchUpdateTableRowsResponse'

-> BatchUpdateTableRowsResponse

Create a value of BatchUpdateTableRowsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedBatchItems:BatchUpdateTableRowsResponse', batchUpdateTableRowsResponse_failedBatchItems - The list of batch items in the request that could not be updated in the table. Each element in this list contains one item from the request that could not be updated in the table along with the reason why that item could not be updated.

$sel:httpStatus:BatchUpdateTableRowsResponse', batchUpdateTableRowsResponse_httpStatus - The response's http status code.

$sel:workbookCursor:BatchUpdateTableRowsResponse', batchUpdateTableRowsResponse_workbookCursor - The updated workbook cursor after adding the new rows at the end of the table.

Response Lenses

batchUpdateTableRowsResponse_failedBatchItems :: Lens' BatchUpdateTableRowsResponse (Maybe [FailedBatchItem])

The list of batch items in the request that could not be updated in the table. Each element in this list contains one item from the request that could not be updated in the table along with the reason why that item could not be updated.

batchUpdateTableRowsResponse_httpStatus :: Lens' BatchUpdateTableRowsResponse Int

The response's http status code.

batchUpdateTableRowsResponse_workbookCursor :: Lens' BatchUpdateTableRowsResponse Integer

The updated workbook cursor after adding the new rows at the end of the table.