Network.AWS.HoneyCode.GetScreenData

Description

The GetScreenData API allows retrieval of data from a screen in a Honeycode app. The API allows setting local variables in the screen to filter, sort or otherwise affect what will be displayed on the screen.

Creating a Request

data GetScreenData

See: newGetScreenData smart constructor.

Constructors

GetScreenData'

Fields

variables :: Maybe (Sensitive (HashMap Text (Sensitive VariableValue)))

Variables are optional and are needed only if the screen requires them to render correctly. Variables are specified as a map where the key is the name of the variable as defined on the screen. The value is an object which currently has only one property, rawValue, which holds the value of the variable to be passed to the screen.

nextToken :: Maybe Text

This parameter is optional. If a nextToken is not specified, the API returns the first page of data.

Pagination tokens expire after 1 hour. If you use a token that was returned more than an hour back, the API will throw ValidationException.

maxResults :: Maybe Natural

The number of results to be returned on a single page. Specify a number between 1 and 100. The maximum value is 100.

This parameter is optional. If you don't specify this parameter, the default page size is 100.

workbookId :: Text

The ID of the workbook that contains the screen.

appId :: Text

The ID of the app that contains the screem.

screenId :: Text

The ID of the screen.

newGetScreenData

Arguments

:: Text
$sel:workbookId:GetScreenData'

-> Text
$sel:appId:GetScreenData'

-> Text
$sel:screenId:GetScreenData'

-> GetScreenData

Create a value of GetScreenData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:variables:GetScreenData', getScreenData_variables - Variables are optional and are needed only if the screen requires them to render correctly. Variables are specified as a map where the key is the name of the variable as defined on the screen. The value is an object which currently has only one property, rawValue, which holds the value of the variable to be passed to the screen.

$sel:nextToken:GetScreenData', getScreenData_nextToken - This parameter is optional. If a nextToken is not specified, the API returns the first page of data.

Pagination tokens expire after 1 hour. If you use a token that was returned more than an hour back, the API will throw ValidationException.

$sel:maxResults:GetScreenData', getScreenData_maxResults - The number of results to be returned on a single page. Specify a number between 1 and 100. The maximum value is 100.

This parameter is optional. If you don't specify this parameter, the default page size is 100.

$sel:workbookId:GetScreenData', getScreenData_workbookId - The ID of the workbook that contains the screen.

$sel:appId:GetScreenData', getScreenData_appId - The ID of the app that contains the screem.

$sel:screenId:GetScreenData', getScreenData_screenId - The ID of the screen.

Request Lenses

getScreenData_variables :: Lens' GetScreenData (Maybe (HashMap Text VariableValue))

Variables are optional and are needed only if the screen requires them to render correctly. Variables are specified as a map where the key is the name of the variable as defined on the screen. The value is an object which currently has only one property, rawValue, which holds the value of the variable to be passed to the screen.

getScreenData_nextToken :: Lens' GetScreenData (Maybe Text)

This parameter is optional. If a nextToken is not specified, the API returns the first page of data.

Pagination tokens expire after 1 hour. If you use a token that was returned more than an hour back, the API will throw ValidationException.

getScreenData_maxResults :: Lens' GetScreenData (Maybe Natural)

The number of results to be returned on a single page. Specify a number between 1 and 100. The maximum value is 100.

This parameter is optional. If you don't specify this parameter, the default page size is 100.

getScreenData_workbookId :: Lens' GetScreenData Text

The ID of the workbook that contains the screen.

getScreenData_appId :: Lens' GetScreenData Text

The ID of the app that contains the screem.

getScreenData_screenId :: Lens' GetScreenData Text

The ID of the screen.

Destructuring the Response

data GetScreenDataResponse

See: newGetScreenDataResponse smart constructor.

Constructors

GetScreenDataResponse'

Fields

nextToken :: Maybe Text

Provides the pagination token to load the next page if there are more results matching the request. If a pagination token is not present in the response, it means that all data matching the query has been loaded.

httpStatus :: Int

The response's http status code.

results :: HashMap Text ResultSet

A map of all the rows on the screen keyed by block name.

workbookCursor :: Integer

Indicates the cursor of the workbook at which the data returned by this workbook is read. Workbook cursor keeps increasing with every update and the increments are not sequential.

newGetScreenDataResponse

Arguments

:: Int
$sel:httpStatus:GetScreenDataResponse'

-> Integer
$sel:workbookCursor:GetScreenDataResponse'

-> GetScreenDataResponse

Create a value of GetScreenDataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetScreenData', getScreenDataResponse_nextToken - Provides the pagination token to load the next page if there are more results matching the request. If a pagination token is not present in the response, it means that all data matching the query has been loaded.

$sel:httpStatus:GetScreenDataResponse', getScreenDataResponse_httpStatus - The response's http status code.

$sel:results:GetScreenDataResponse', getScreenDataResponse_results - A map of all the rows on the screen keyed by block name.

$sel:workbookCursor:GetScreenDataResponse', getScreenDataResponse_workbookCursor - Indicates the cursor of the workbook at which the data returned by this workbook is read. Workbook cursor keeps increasing with every update and the increments are not sequential.

Response Lenses

getScreenDataResponse_nextToken :: Lens' GetScreenDataResponse (Maybe Text)

Provides the pagination token to load the next page if there are more results matching the request. If a pagination token is not present in the response, it means that all data matching the query has been loaded.

getScreenDataResponse_httpStatus :: Lens' GetScreenDataResponse Int

The response's http status code.

getScreenDataResponse_results :: Lens' GetScreenDataResponse (HashMap Text ResultSet)

A map of all the rows on the screen keyed by block name.

getScreenDataResponse_workbookCursor :: Lens' GetScreenDataResponse Integer

Indicates the cursor of the workbook at which the data returned by this workbook is read. Workbook cursor keeps increasing with every update and the increments are not sequential.