Network.AWS.HoneyCode.Lens

Description

Operations

BatchDeleteTableRows

batchDeleteTableRows_clientRequestToken :: Lens' BatchDeleteTableRows (Maybe Text)

The request token for performing the delete action. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will not perform the action again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

batchDeleteTableRows_workbookId :: Lens' BatchDeleteTableRows Text

The ID of the workbook where the rows are being deleted.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

batchDeleteTableRows_tableId :: Lens' BatchDeleteTableRows Text

The ID of the table where the rows are being deleted.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

batchDeleteTableRows_rowIds :: Lens' BatchDeleteTableRows (NonEmpty Text)

The list of row ids to delete from the table. You need to specify at least one row id in this list.

Note that if one of the row ids provided in the request does not exist in the table, then the request fails and no rows are deleted from the table.

batchDeleteTableRowsResponse_failedBatchItems :: Lens' BatchDeleteTableRowsResponse (Maybe [FailedBatchItem])

The list of row ids in the request that could not be deleted from the table. Each element in this list contains one row id from the request that could not be deleted along with the reason why that item could not be deleted.

batchDeleteTableRowsResponse_httpStatus :: Lens' BatchDeleteTableRowsResponse Int

The response's http status code.

batchDeleteTableRowsResponse_workbookCursor :: Lens' BatchDeleteTableRowsResponse Integer

The updated workbook cursor after deleting the rows from the table.

BatchUpdateTableRows

batchUpdateTableRows_clientRequestToken :: Lens' BatchUpdateTableRows (Maybe Text)

The request token for performing the update action. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will not perform the action again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

batchUpdateTableRows_workbookId :: Lens' BatchUpdateTableRows Text

The ID of the workbook where the rows are being updated.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

batchUpdateTableRows_tableId :: Lens' BatchUpdateTableRows Text

The ID of the table where the rows are being updated.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

batchUpdateTableRows_rowsToUpdate :: Lens' BatchUpdateTableRows (NonEmpty UpdateRowData)

The list of rows to update in the table. Each item in this list needs to contain the row id to update along with the map of column id to cell values for each column in that row that needs to be updated. You need to specify at least one row in this list, and for each row, you need to specify at least one column to update.

Note that if one of the row or column ids in the request does not exist in the table, then the request fails and no updates are made to the table.

batchUpdateTableRowsResponse_failedBatchItems :: Lens' BatchUpdateTableRowsResponse (Maybe [FailedBatchItem])

The list of batch items in the request that could not be updated in the table. Each element in this list contains one item from the request that could not be updated in the table along with the reason why that item could not be updated.

batchUpdateTableRowsResponse_httpStatus :: Lens' BatchUpdateTableRowsResponse Int

The response's http status code.

batchUpdateTableRowsResponse_workbookCursor :: Lens' BatchUpdateTableRowsResponse Integer

The updated workbook cursor after adding the new rows at the end of the table.

ListTableRows

listTableRows_rowIds :: Lens' ListTableRows (Maybe (NonEmpty Text))

This parameter is optional. If one or more row ids are specified in this list, then only the specified row ids are returned in the result. If no row ids are specified here, then all the rows in the table are returned.

listTableRows_nextToken :: Lens' ListTableRows (Maybe Text)

This parameter is optional. If a nextToken is not specified, the API returns the first page of data.

Pagination tokens expire after 1 hour. If you use a token that was returned more than an hour back, the API will throw ValidationException.

listTableRows_maxResults :: Lens' ListTableRows (Maybe Natural)

The maximum number of rows to return in each page of the results.

listTableRows_workbookId :: Lens' ListTableRows Text

The ID of the workbook that contains the table whose rows are being retrieved.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

listTableRows_tableId :: Lens' ListTableRows Text

The ID of the table whose rows are being retrieved.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

listTableRowsResponse_rowIdsNotFound :: Lens' ListTableRowsResponse (Maybe (NonEmpty Text))

The list of row ids included in the request that were not found in the table.

listTableRowsResponse_nextToken :: Lens' ListTableRowsResponse (Maybe Text)

Provides the pagination token to load the next page if there are more results matching the request. If a pagination token is not present in the response, it means that all data matching the request has been loaded.

listTableRowsResponse_httpStatus :: Lens' ListTableRowsResponse Int

The response's http status code.

listTableRowsResponse_columnIds :: Lens' ListTableRowsResponse (NonEmpty Text)

The list of columns in the table whose row data is returned in the result.

listTableRowsResponse_rows :: Lens' ListTableRowsResponse [TableRow]

The list of rows in the table. Note that this result is paginated, so this list contains a maximum of 100 rows.

listTableRowsResponse_workbookCursor :: Lens' ListTableRowsResponse Integer

Indicates the cursor of the workbook at which the data returned by this request is read. Workbook cursor keeps increasing with every update and the increments are not sequential.

InvokeScreenAutomation

invokeScreenAutomation_rowId :: Lens' InvokeScreenAutomation (Maybe Text)

The row ID for the automation if the automation is defined inside a block with source or list.

invokeScreenAutomation_variables :: Lens' InvokeScreenAutomation (Maybe (HashMap Text VariableValue))

Variables are specified as a map where the key is the name of the variable as defined on the screen. The value is an object which currently has only one property, rawValue, which holds the value of the variable to be passed to the screen. Any variables defined in a screen are required to be passed in the call.

invokeScreenAutomation_clientRequestToken :: Lens' InvokeScreenAutomation (Maybe Text)

The request token for performing the automation action. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will return the response of the previous call rather than performing the action again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

invokeScreenAutomation_workbookId :: Lens' InvokeScreenAutomation Text

The ID of the workbook that contains the screen automation.

invokeScreenAutomation_appId :: Lens' InvokeScreenAutomation Text

The ID of the app that contains the screen automation.

invokeScreenAutomation_screenId :: Lens' InvokeScreenAutomation Text

The ID of the screen that contains the screen automation.

invokeScreenAutomation_screenAutomationId :: Lens' InvokeScreenAutomation Text

The ID of the automation action to be performed.

invokeScreenAutomationResponse_httpStatus :: Lens' InvokeScreenAutomationResponse Int

The response's http status code.

invokeScreenAutomationResponse_workbookCursor :: Lens' InvokeScreenAutomationResponse Integer

The updated workbook cursor after performing the automation action.

DescribeTableDataImportJob

describeTableDataImportJob_workbookId :: Lens' DescribeTableDataImportJob Text

The ID of the workbook into which data was imported.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

describeTableDataImportJob_tableId :: Lens' DescribeTableDataImportJob Text

The ID of the table into which data was imported.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

describeTableDataImportJob_jobId :: Lens' DescribeTableDataImportJob Text

The ID of the job that was returned by the StartTableDataImportJob request.

If a job with the specified id could not be found, this API throws ResourceNotFoundException.

describeTableDataImportJobResponse_httpStatus :: Lens' DescribeTableDataImportJobResponse Int

The response's http status code.

describeTableDataImportJobResponse_jobStatus :: Lens' DescribeTableDataImportJobResponse TableDataImportJobStatus

The current status of the import job.

describeTableDataImportJobResponse_message :: Lens' DescribeTableDataImportJobResponse Text

A message providing more details about the current status of the import job.

describeTableDataImportJobResponse_jobMetadata :: Lens' DescribeTableDataImportJobResponse TableDataImportJobMetadata

The metadata about the job that was submitted for import.

StartTableDataImportJob

startTableDataImportJob_workbookId :: Lens' StartTableDataImportJob Text

The ID of the workbook where the rows are being imported.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

startTableDataImportJob_dataSource :: Lens' StartTableDataImportJob ImportDataSource

The source of the data that is being imported. The size of source must be no larger than 100 MB. Source must have no more than 100,000 cells and no more than 1,000 rows.

startTableDataImportJob_dataFormat :: Lens' StartTableDataImportJob ImportSourceDataFormat

The format of the data that is being imported. Currently the only option supported is "DELIMITED_TEXT".

startTableDataImportJob_destinationTableId :: Lens' StartTableDataImportJob Text

The ID of the table where the rows are being imported.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

startTableDataImportJob_importOptions :: Lens' StartTableDataImportJob ImportOptions

The options for customizing this import request.

startTableDataImportJob_clientRequestToken :: Lens' StartTableDataImportJob Text

The request token for performing the update action. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will not perform the action again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

startTableDataImportJobResponse_httpStatus :: Lens' StartTableDataImportJobResponse Int

The response's http status code.

startTableDataImportJobResponse_jobId :: Lens' StartTableDataImportJobResponse Text

The id that is assigned to this import job. Future requests to find out the status of this import job need to send this id in the appropriate parameter in the request.

startTableDataImportJobResponse_jobStatus :: Lens' StartTableDataImportJobResponse TableDataImportJobStatus

The status of the import job immediately after submitting the request.

BatchCreateTableRows

batchCreateTableRows_clientRequestToken :: Lens' BatchCreateTableRows (Maybe Text)

The request token for performing the batch create operation. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will not perform the operation again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

batchCreateTableRows_workbookId :: Lens' BatchCreateTableRows Text

The ID of the workbook where the new rows are being added.

If a workbook with the specified ID could not be found, this API throws ResourceNotFoundException.

batchCreateTableRows_tableId :: Lens' BatchCreateTableRows Text

The ID of the table where the new rows are being added.

If a table with the specified ID could not be found, this API throws ResourceNotFoundException.

batchCreateTableRows_rowsToCreate :: Lens' BatchCreateTableRows (NonEmpty CreateRowData)

The list of rows to create at the end of the table. Each item in this list needs to have a batch item id to uniquely identify the element in the request and the cells to create for that row. You need to specify at least one item in this list.

Note that if one of the column ids in any of the rows in the request does not exist in the table, then the request fails and no updates are made to the table.

batchCreateTableRowsResponse_failedBatchItems :: Lens' BatchCreateTableRowsResponse (Maybe [FailedBatchItem])

The list of batch items in the request that could not be added to the table. Each element in this list contains one item from the request that could not be added to the table along with the reason why that item could not be added.

batchCreateTableRowsResponse_httpStatus :: Lens' BatchCreateTableRowsResponse Int

The response's http status code.

batchCreateTableRowsResponse_workbookCursor :: Lens' BatchCreateTableRowsResponse Integer

The updated workbook cursor after adding the new rows at the end of the table.

batchCreateTableRowsResponse_createdRows :: Lens' BatchCreateTableRowsResponse (HashMap Text Text)

The map of batch item id to the row id that was created for that item.

ListTables

listTables_nextToken :: Lens' ListTables (Maybe Text)

This parameter is optional. If a nextToken is not specified, the API returns the first page of data.

Pagination tokens expire after 1 hour. If you use a token that was returned more than an hour back, the API will throw ValidationException.

listTables_maxResults :: Lens' ListTables (Maybe Natural)

The maximum number of tables to return in each page of the results.

listTables_workbookId :: Lens' ListTables Text

The ID of the workbook whose tables are being retrieved.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

listTablesResponse_nextToken :: Lens' ListTablesResponse (Maybe Text)

Provides the pagination token to load the next page if there are more results matching the request. If a pagination token is not present in the response, it means that all data matching the request has been loaded.

listTablesResponse_workbookCursor :: Lens' ListTablesResponse (Maybe Integer)

Indicates the cursor of the workbook at which the data returned by this request is read. Workbook cursor keeps increasing with every update and the increments are not sequential.

listTablesResponse_httpStatus :: Lens' ListTablesResponse Int

The response's http status code.

listTablesResponse_tables :: Lens' ListTablesResponse [Table]

The list of tables in the workbook.

GetScreenData

getScreenData_variables :: Lens' GetScreenData (Maybe (HashMap Text VariableValue))

Variables are optional and are needed only if the screen requires them to render correctly. Variables are specified as a map where the key is the name of the variable as defined on the screen. The value is an object which currently has only one property, rawValue, which holds the value of the variable to be passed to the screen.

getScreenData_nextToken :: Lens' GetScreenData (Maybe Text)

This parameter is optional. If a nextToken is not specified, the API returns the first page of data.

Pagination tokens expire after 1 hour. If you use a token that was returned more than an hour back, the API will throw ValidationException.

getScreenData_maxResults :: Lens' GetScreenData (Maybe Natural)

The number of results to be returned on a single page. Specify a number between 1 and 100. The maximum value is 100.

This parameter is optional. If you don't specify this parameter, the default page size is 100.

getScreenData_workbookId :: Lens' GetScreenData Text

The ID of the workbook that contains the screen.

getScreenData_appId :: Lens' GetScreenData Text

The ID of the app that contains the screem.

getScreenData_screenId :: Lens' GetScreenData Text

The ID of the screen.

getScreenDataResponse_nextToken :: Lens' GetScreenDataResponse (Maybe Text)

Provides the pagination token to load the next page if there are more results matching the request. If a pagination token is not present in the response, it means that all data matching the query has been loaded.

getScreenDataResponse_httpStatus :: Lens' GetScreenDataResponse Int

The response's http status code.

getScreenDataResponse_results :: Lens' GetScreenDataResponse (HashMap Text ResultSet)

A map of all the rows on the screen keyed by block name.

getScreenDataResponse_workbookCursor :: Lens' GetScreenDataResponse Integer

Indicates the cursor of the workbook at which the data returned by this workbook is read. Workbook cursor keeps increasing with every update and the increments are not sequential.

QueryTableRows

queryTableRows_nextToken :: Lens' QueryTableRows (Maybe Text)

This parameter is optional. If a nextToken is not specified, the API returns the first page of data.

Pagination tokens expire after 1 hour. If you use a token that was returned more than an hour back, the API will throw ValidationException.

queryTableRows_maxResults :: Lens' QueryTableRows (Maybe Natural)

The maximum number of rows to return in each page of the results.

queryTableRows_workbookId :: Lens' QueryTableRows Text

The ID of the workbook whose table rows are being queried.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

queryTableRows_tableId :: Lens' QueryTableRows Text

The ID of the table whose rows are being queried.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

queryTableRows_filterFormula :: Lens' QueryTableRows Filter

An object that represents a filter formula along with the id of the context row under which the filter function needs to evaluate.

queryTableRowsResponse_nextToken :: Lens' QueryTableRowsResponse (Maybe Text)

Provides the pagination token to load the next page if there are more results matching the request. If a pagination token is not present in the response, it means that all data matching the request has been loaded.

queryTableRowsResponse_httpStatus :: Lens' QueryTableRowsResponse Int

The response's http status code.

queryTableRowsResponse_columnIds :: Lens' QueryTableRowsResponse (NonEmpty Text)

The list of columns in the table whose row data is returned in the result.

queryTableRowsResponse_rows :: Lens' QueryTableRowsResponse [TableRow]

The list of rows in the table that match the query filter.

queryTableRowsResponse_workbookCursor :: Lens' QueryTableRowsResponse Integer

Indicates the cursor of the workbook at which the data returned by this request is read. Workbook cursor keeps increasing with every update and the increments are not sequential.

BatchUpsertTableRows

batchUpsertTableRows_clientRequestToken :: Lens' BatchUpsertTableRows (Maybe Text)

The request token for performing the update action. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will not perform the action again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

batchUpsertTableRows_workbookId :: Lens' BatchUpsertTableRows Text

The ID of the workbook where the rows are being upserted.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

batchUpsertTableRows_tableId :: Lens' BatchUpsertTableRows Text

The ID of the table where the rows are being upserted.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

batchUpsertTableRows_rowsToUpsert :: Lens' BatchUpsertTableRows [UpsertRowData]

The list of rows to upsert in the table. Each item in this list needs to have a batch item id to uniquely identify the element in the request, a filter expression to find the rows to update for that element and the cell values to set for each column in the upserted rows. You need to specify at least one item in this list.

Note that if one of the filter formulas in the request fails to evaluate because of an error or one of the column ids in any of the rows does not exist in the table, then the request fails and no updates are made to the table.

batchUpsertTableRowsResponse_failedBatchItems :: Lens' BatchUpsertTableRowsResponse (Maybe [FailedBatchItem])

The list of batch items in the request that could not be updated or appended in the table. Each element in this list contains one item from the request that could not be updated in the table along with the reason why that item could not be updated or appended.

batchUpsertTableRowsResponse_httpStatus :: Lens' BatchUpsertTableRowsResponse Int

The response's http status code.

batchUpsertTableRowsResponse_rows :: Lens' BatchUpsertTableRowsResponse (HashMap Text UpsertRowsResult)

A map with the batch item id as the key and the result of the upsert operation as the value. The result of the upsert operation specifies whether existing rows were updated or a new row was appended, along with the list of row ids that were affected.

batchUpsertTableRowsResponse_workbookCursor :: Lens' BatchUpsertTableRowsResponse Integer

The updated workbook cursor after updating or appending rows in the table.

ListTableColumns

listTableColumns_nextToken :: Lens' ListTableColumns (Maybe Text)

This parameter is optional. If a nextToken is not specified, the API returns the first page of data.

Pagination tokens expire after 1 hour. If you use a token that was returned more than an hour back, the API will throw ValidationException.

listTableColumns_workbookId :: Lens' ListTableColumns Text

The ID of the workbook that contains the table whose columns are being retrieved.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

listTableColumns_tableId :: Lens' ListTableColumns Text

The ID of the table whose columns are being retrieved.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

listTableColumnsResponse_nextToken :: Lens' ListTableColumnsResponse (Maybe Text)

Provides the pagination token to load the next page if there are more results matching the request. If a pagination token is not present in the response, it means that all data matching the request has been loaded.

listTableColumnsResponse_workbookCursor :: Lens' ListTableColumnsResponse (Maybe Integer)

Indicates the cursor of the workbook at which the data returned by this request is read. Workbook cursor keeps increasing with every update and the increments are not sequential.

listTableColumnsResponse_httpStatus :: Lens' ListTableColumnsResponse Int

The response's http status code.

listTableColumnsResponse_tableColumns :: Lens' ListTableColumnsResponse [TableColumn]

The list of columns in the table.

Types

Cell

cell_rawValue :: Lens' Cell (Maybe Text)

The raw value of the data contained in the cell. The raw value depends on the format of the data in the cell. However the attribute in the API return value is always a string containing the raw value.

Cells with format DATE, DATE_TIME or TIME have the raw value as a floating point number where the whole number represents the number of days since 1/1/1900 and the fractional part represents the fraction of the day since midnight. For example, a cell with date 11/3/2020 has the raw value "44138". A cell with the time 9:00 AM has the raw value "0.375" and a cell with date/time value of 11/3/2020 9:00 AM has the raw value "44138.375". Notice that even though the raw value is a number in all three cases, it is still represented as a string.

Cells with format NUMBER, CURRENCY, PERCENTAGE and ACCOUNTING have the raw value of the data as the number representing the data being displayed. For example, the number 1.325 with two decimal places in the format will have it's raw value as "1.325" and formatted value as "1.33". A currency value for $10 will have the raw value as "10" and formatted value as "$10.00". A value representing 20% with two decimal places in the format will have its raw value as "0.2" and the formatted value as "20.00%". An accounting value of -$25 will have "-25" as the raw value and "$ (25.00)" as the formatted value.

Cells with format TEXT will have the raw text as the raw value. For example, a cell with text "John Smith" will have "John Smith" as both the raw value and the formatted value.

Cells with format CONTACT will have the name of the contact as a formatted value and the email address of the contact as the raw value. For example, a contact for John Smith will have "John Smith" as the formatted value and "john.smith@example.com" as the raw value.

Cells with format ROWLINK (aka picklist) will have the first column of the linked row as the formatted value and the row id of the linked row as the raw value. For example, a cell containing a picklist to a table that displays task status might have "Completed" as the formatted value and "row:dfcefaee-5b37-4355-8f28-40c3e4ff5dd4/ca432b2f-b8eb-431d-9fb5-cbe0342f9f03" as the raw value.

Cells with format AUTO or cells without any format that are auto-detected as one of the formats above will contain the raw and formatted values as mentioned above, based on the auto-detected formats. If there is no auto-detected format, the raw and formatted values will be the same as the data in the cell.

cell_format :: Lens' Cell (Maybe Format)

The format of the cell. If this field is empty, then the format is either not specified in the workbook or the format is set to AUTO.

cell_formula :: Lens' Cell (Maybe Text)

The formula contained in the cell. This field is empty if a cell does not have a formula.

cell_formattedValue :: Lens' Cell (Maybe Text)

The formatted value of the cell. This is the value that you see displayed in the cell in the UI.

Note that the formatted value of a cell is always represented as a string irrespective of the data that is stored in the cell. For example, if a cell contains a date, the formatted value of the cell is the string representation of the formatted date being shown in the cell in the UI. See details in the rawValue field below for how cells of different formats will have different raw and formatted values.

CellInput

cellInput_fact :: Lens' CellInput (Maybe Text)

Fact represents the data that is entered into a cell. This data can be free text or a formula. Formulas need to start with the equals (=) sign.

ColumnMetadata

columnMetadata_name :: Lens' ColumnMetadata Text

The name of the column.

columnMetadata_format :: Lens' ColumnMetadata Format

The format of the column.

CreateRowData

createRowData_batchItemId :: Lens' CreateRowData Text

An external identifier that represents the single row that is being created as part of the BatchCreateTableRows request. This can be any string that you can use to identify the row in the request. The BatchCreateTableRows API puts the batch item id in the results to allow you to link data in the request to data in the results.

createRowData_cellsToCreate :: Lens' CreateRowData (HashMap Text CellInput)

A map representing the cells to create in the new row. The key is the column id of the cell and the value is the CellInput object that represents the data to set in that cell.

DataItem

dataItem_rawValue :: Lens' DataItem (Maybe Text)

The raw value of the data. e.g. jsmith@example.com

dataItem_overrideFormat :: Lens' DataItem (Maybe Format)

The overrideFormat is optional and is specified only if a particular row of data has a different format for the data than the default format defined on the screen or the table.

dataItem_formattedValue :: Lens' DataItem (Maybe Text)

The formatted value of the data. e.g. John Smith.

DelimitedTextImportOptions

delimitedTextImportOptions_ignoreEmptyRows :: Lens' DelimitedTextImportOptions (Maybe Bool)

A parameter to indicate whether empty rows should be ignored or be included in the import.

delimitedTextImportOptions_hasHeaderRow :: Lens' DelimitedTextImportOptions (Maybe Bool)

Indicates whether the input file has a header row at the top containing the column names.

delimitedTextImportOptions_dataCharacterEncoding :: Lens' DelimitedTextImportOptions (Maybe ImportDataCharacterEncoding)

The encoding of the data in the input file.

delimitedTextImportOptions_delimiter :: Lens' DelimitedTextImportOptions Text

The delimiter to use for separating columns in a single row of the input.

DestinationOptions

destinationOptions_columnMap :: Lens' DestinationOptions (Maybe (HashMap Text SourceDataColumnProperties))

A map of the column id to the import properties for each column.

FailedBatchItem

failedBatchItem_id :: Lens' FailedBatchItem Text

The id of the batch item that failed. This is the batch item id for the BatchCreateTableRows and BatchUpsertTableRows operations and the row id for the BatchUpdateTableRows and BatchDeleteTableRows operations.

failedBatchItem_errorMessage :: Lens' FailedBatchItem Text

The error message that indicates why the batch item failed.

Filter

filter_contextRowId :: Lens' Filter (Maybe Text)

The optional contextRowId attribute can be used to specify the row id of the context row if the filter formula contains unqualified references to table columns and needs a context row to evaluate them successfully.

filter_formula :: Lens' Filter Text

A formula representing a filter function that returns zero or more matching rows from a table. Valid formulas in this field return a list of rows from a table. The most common ways of writing a formula to return a list of rows are to use the FindRow() or Filter() functions. Any other formula that returns zero or more rows is also acceptable. For example, you can use a formula that points to a cell that contains a filter function.

ImportDataSource

importDataSource_dataSourceConfig :: Lens' ImportDataSource ImportDataSourceConfig

The configuration parameters for the data source of the import

ImportDataSourceConfig

importDataSourceConfig_dataSourceUrl :: Lens' ImportDataSourceConfig (Maybe Text)

The URL from which source data will be downloaded for the import request.

ImportJobSubmitter

importJobSubmitter_email :: Lens' ImportJobSubmitter (Maybe Text)

The email id of the submitter of the import job, if available.

importJobSubmitter_userArn :: Lens' ImportJobSubmitter (Maybe Text)

The AWS user ARN of the submitter of the import job, if available.

ImportOptions

importOptions_delimitedTextOptions :: Lens' ImportOptions (Maybe DelimitedTextImportOptions)

Options relating to parsing delimited text. Required if dataFormat is DELIMITED_TEXT.

importOptions_destinationOptions :: Lens' ImportOptions (Maybe DestinationOptions)

Options relating to the destination of the import request.

ResultRow

resultRow_rowId :: Lens' ResultRow (Maybe Text)

The ID for a particular row.

resultRow_dataItems :: Lens' ResultRow [DataItem]

List of all the data cells in a row.

ResultSet

resultSet_headers :: Lens' ResultSet [ColumnMetadata]

List of headers for all the data cells in the block. The header identifies the name and default format of the data cell. Data cells appear in the same order in all rows as defined in the header. The names and formats are not repeated in the rows. If a particular row does not have a value for a data cell, a blank value is used.

For example, a task list that displays the task name, due date and assigned person might have headers [ { "name": "Task Name"}, {"name": "Due Date", "format": "DATE"}, {"name": "Assigned", "format": "CONTACT"} ]. Every row in the result will have the task name as the first item, due date as the second item and assigned person as the third item. If a particular task does not have a due date, that row will still have a blank value in the second element and the assigned person will still be in the third element.

resultSet_rows :: Lens' ResultSet [ResultRow]

List of rows returned by the request. Each row has a row Id and a list of data cells in that row. The data cells will be present in the same order as they are defined in the header.

SourceDataColumnProperties

sourceDataColumnProperties_columnIndex :: Lens' SourceDataColumnProperties (Maybe Natural)

The index of the column in the input file.

Table

table_tableId :: Lens' Table (Maybe Text)

The id of the table.

table_tableName :: Lens' Table (Maybe Text)

The name of the table.

TableColumn

tableColumn_format :: Lens' TableColumn (Maybe Format)

The column level format that is applied in the table. An empty value in this field means that the column format is the default value 'AUTO'.

tableColumn_tableColumnName :: Lens' TableColumn (Maybe Text)

The name of the column in the table.

tableColumn_tableColumnId :: Lens' TableColumn (Maybe Text)

The id of the column in the table.

TableDataImportJobMetadata

tableDataImportJobMetadata_submitter :: Lens' TableDataImportJobMetadata ImportJobSubmitter

Details about the submitter of the import request.

tableDataImportJobMetadata_submitTime :: Lens' TableDataImportJobMetadata UTCTime

The timestamp when the job was submitted for import.

tableDataImportJobMetadata_importOptions :: Lens' TableDataImportJobMetadata ImportOptions

The options that was specified at the time of submitting the import request.

tableDataImportJobMetadata_dataSource :: Lens' TableDataImportJobMetadata ImportDataSource

The source of the data that was submitted for import.

TableRow

tableRow_rowId :: Lens' TableRow Text

The id of the row in the table.

tableRow_cells :: Lens' TableRow [Cell]

A list of cells in the table row. The cells appear in the same order as the columns of the table.

UpdateRowData

updateRowData_rowId :: Lens' UpdateRowData Text

The id of the row that needs to be updated.

updateRowData_cellsToUpdate :: Lens' UpdateRowData (HashMap Text CellInput)

A map representing the cells to update in the given row. The key is the column id of the cell and the value is the CellInput object that represents the data to set in that cell.

UpsertRowData

upsertRowData_batchItemId :: Lens' UpsertRowData Text

An external identifier that represents a single item in the request that is being upserted as part of the BatchUpsertTableRows request. This can be any string that you can use to identify the item in the request. The BatchUpsertTableRows API puts the batch item id in the results to allow you to link data in the request to data in the results.

upsertRowData_filter :: Lens' UpsertRowData Filter

The filter formula to use to find existing matching rows to update. The formula needs to return zero or more rows. If the formula returns 0 rows, then a new row will be appended in the target table. If the formula returns one or more rows, then the returned rows will be updated.

Note that the filter formula needs to return rows from the target table for the upsert operation to succeed. If the filter formula has a syntax error or it doesn't evaluate to zero or more rows in the target table for any one item in the input list, then the entire BatchUpsertTableRows request fails and no updates are made to the table.

upsertRowData_cellsToUpdate :: Lens' UpsertRowData (HashMap Text CellInput)

A map representing the cells to update for the matching rows or an appended row. The key is the column id of the cell and the value is the CellInput object that represents the data to set in that cell.

UpsertRowsResult

upsertRowsResult_rowIds :: Lens' UpsertRowsResult (NonEmpty Text)

The list of row ids that were changed as part of an upsert row operation. If the upsert resulted in an update, this list could potentially contain multiple rows that matched the filter and hence got updated. If the upsert resulted in an append, this list would only have the single row that was appended.

upsertRowsResult_upsertAction :: Lens' UpsertRowsResult UpsertAction

The result of the upsert action.

VariableValue

variableValue_rawValue :: Lens' VariableValue Text

Raw value of the variable.