Network.AWS.HoneyCode.ListTableRows

Description

The ListTableRows API allows you to retrieve a list of all the rows in a table in a workbook.

This operation returns paginated results.

Creating a Request

data ListTableRows

See: newListTableRows smart constructor.

Constructors

ListTableRows'

Fields

rowIds :: Maybe (NonEmpty Text)

This parameter is optional. If one or more row ids are specified in this list, then only the specified row ids are returned in the result. If no row ids are specified here, then all the rows in the table are returned.

nextToken :: Maybe Text

This parameter is optional. If a nextToken is not specified, the API returns the first page of data.

Pagination tokens expire after 1 hour. If you use a token that was returned more than an hour back, the API will throw ValidationException.

maxResults :: Maybe Natural

The maximum number of rows to return in each page of the results.

workbookId :: Text

The ID of the workbook that contains the table whose rows are being retrieved.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

tableId :: Text

The ID of the table whose rows are being retrieved.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

newListTableRows

Arguments

:: Text
$sel:workbookId:ListTableRows'

-> Text
$sel:tableId:ListTableRows'

-> ListTableRows

Create a value of ListTableRows with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rowIds:ListTableRows', listTableRows_rowIds - This parameter is optional. If one or more row ids are specified in this list, then only the specified row ids are returned in the result. If no row ids are specified here, then all the rows in the table are returned.

$sel:nextToken:ListTableRows', listTableRows_nextToken - This parameter is optional. If a nextToken is not specified, the API returns the first page of data.

Pagination tokens expire after 1 hour. If you use a token that was returned more than an hour back, the API will throw ValidationException.

$sel:maxResults:ListTableRows', listTableRows_maxResults - The maximum number of rows to return in each page of the results.

$sel:workbookId:ListTableRows', listTableRows_workbookId - The ID of the workbook that contains the table whose rows are being retrieved.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

$sel:tableId:ListTableRows', listTableRows_tableId - The ID of the table whose rows are being retrieved.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

Request Lenses

listTableRows_rowIds :: Lens' ListTableRows (Maybe (NonEmpty Text))

This parameter is optional. If one or more row ids are specified in this list, then only the specified row ids are returned in the result. If no row ids are specified here, then all the rows in the table are returned.

listTableRows_nextToken :: Lens' ListTableRows (Maybe Text)

This parameter is optional. If a nextToken is not specified, the API returns the first page of data.

Pagination tokens expire after 1 hour. If you use a token that was returned more than an hour back, the API will throw ValidationException.

listTableRows_maxResults :: Lens' ListTableRows (Maybe Natural)

The maximum number of rows to return in each page of the results.

listTableRows_workbookId :: Lens' ListTableRows Text

The ID of the workbook that contains the table whose rows are being retrieved.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

listTableRows_tableId :: Lens' ListTableRows Text

The ID of the table whose rows are being retrieved.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

Destructuring the Response

data ListTableRowsResponse

See: newListTableRowsResponse smart constructor.

Constructors

ListTableRowsResponse'

Fields

rowIdsNotFound :: Maybe (NonEmpty Text)

The list of row ids included in the request that were not found in the table.

nextToken :: Maybe Text

Provides the pagination token to load the next page if there are more results matching the request. If a pagination token is not present in the response, it means that all data matching the request has been loaded.

httpStatus :: Int

The response's http status code.

columnIds :: NonEmpty Text

The list of columns in the table whose row data is returned in the result.

rows :: [TableRow]

The list of rows in the table. Note that this result is paginated, so this list contains a maximum of 100 rows.

workbookCursor :: Integer

Indicates the cursor of the workbook at which the data returned by this request is read. Workbook cursor keeps increasing with every update and the increments are not sequential.

newListTableRowsResponse

Arguments

:: Int
$sel:httpStatus:ListTableRowsResponse'

-> NonEmpty Text
$sel:columnIds:ListTableRowsResponse'

-> Integer
$sel:workbookCursor:ListTableRowsResponse'

-> ListTableRowsResponse

Create a value of ListTableRowsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rowIdsNotFound:ListTableRowsResponse', listTableRowsResponse_rowIdsNotFound - The list of row ids included in the request that were not found in the table.

$sel:nextToken:ListTableRows', listTableRowsResponse_nextToken - Provides the pagination token to load the next page if there are more results matching the request. If a pagination token is not present in the response, it means that all data matching the request has been loaded.

$sel:httpStatus:ListTableRowsResponse', listTableRowsResponse_httpStatus - The response's http status code.

$sel:columnIds:ListTableRowsResponse', listTableRowsResponse_columnIds - The list of columns in the table whose row data is returned in the result.

$sel:rows:ListTableRowsResponse', listTableRowsResponse_rows - The list of rows in the table. Note that this result is paginated, so this list contains a maximum of 100 rows.

$sel:workbookCursor:ListTableRowsResponse', listTableRowsResponse_workbookCursor - Indicates the cursor of the workbook at which the data returned by this request is read. Workbook cursor keeps increasing with every update and the increments are not sequential.

Response Lenses

listTableRowsResponse_rowIdsNotFound :: Lens' ListTableRowsResponse (Maybe (NonEmpty Text))

The list of row ids included in the request that were not found in the table.

listTableRowsResponse_nextToken :: Lens' ListTableRowsResponse (Maybe Text)

Provides the pagination token to load the next page if there are more results matching the request. If a pagination token is not present in the response, it means that all data matching the request has been loaded.

listTableRowsResponse_httpStatus :: Lens' ListTableRowsResponse Int

The response's http status code.

listTableRowsResponse_columnIds :: Lens' ListTableRowsResponse (NonEmpty Text)

The list of columns in the table whose row data is returned in the result.

listTableRowsResponse_rows :: Lens' ListTableRowsResponse [TableRow]

The list of rows in the table. Note that this result is paginated, so this list contains a maximum of 100 rows.

listTableRowsResponse_workbookCursor :: Lens' ListTableRowsResponse Integer

Indicates the cursor of the workbook at which the data returned by this request is read. Workbook cursor keeps increasing with every update and the increments are not sequential.