Network.AWS.HoneyCode.StartTableDataImportJob

Description

The StartTableDataImportJob API allows you to start an import job on a table. This API will only return the id of the job that was started. To find out the status of the import request, you need to call the DescribeTableDataImportJob API.

Creating a Request

data StartTableDataImportJob

See: newStartTableDataImportJob smart constructor.

Constructors

StartTableDataImportJob'

Fields

workbookId :: Text

The ID of the workbook where the rows are being imported.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

dataSource :: ImportDataSource

The source of the data that is being imported. The size of source must be no larger than 100 MB. Source must have no more than 100,000 cells and no more than 1,000 rows.

dataFormat :: ImportSourceDataFormat

The format of the data that is being imported. Currently the only option supported is "DELIMITED_TEXT".

destinationTableId :: Text

The ID of the table where the rows are being imported.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

importOptions :: ImportOptions

The options for customizing this import request.

clientRequestToken :: Text

The request token for performing the update action. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will not perform the action again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

newStartTableDataImportJob

Arguments

:: Text
$sel:workbookId:StartTableDataImportJob'

-> ImportDataSource
$sel:dataSource:StartTableDataImportJob'

-> ImportSourceDataFormat
$sel:dataFormat:StartTableDataImportJob'

-> Text
$sel:destinationTableId:StartTableDataImportJob'

-> ImportOptions
$sel:importOptions:StartTableDataImportJob'

-> Text
$sel:clientRequestToken:StartTableDataImportJob'

-> StartTableDataImportJob

Create a value of StartTableDataImportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workbookId:StartTableDataImportJob', startTableDataImportJob_workbookId - The ID of the workbook where the rows are being imported.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

$sel:dataSource:StartTableDataImportJob', startTableDataImportJob_dataSource - The source of the data that is being imported. The size of source must be no larger than 100 MB. Source must have no more than 100,000 cells and no more than 1,000 rows.

$sel:dataFormat:StartTableDataImportJob', startTableDataImportJob_dataFormat - The format of the data that is being imported. Currently the only option supported is "DELIMITED_TEXT".

$sel:destinationTableId:StartTableDataImportJob', startTableDataImportJob_destinationTableId - The ID of the table where the rows are being imported.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

$sel:importOptions:StartTableDataImportJob', startTableDataImportJob_importOptions - The options for customizing this import request.

$sel:clientRequestToken:StartTableDataImportJob', startTableDataImportJob_clientRequestToken - The request token for performing the update action. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will not perform the action again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

Request Lenses

startTableDataImportJob_workbookId :: Lens' StartTableDataImportJob Text

The ID of the workbook where the rows are being imported.

If a workbook with the specified id could not be found, this API throws ResourceNotFoundException.

startTableDataImportJob_dataSource :: Lens' StartTableDataImportJob ImportDataSource

The source of the data that is being imported. The size of source must be no larger than 100 MB. Source must have no more than 100,000 cells and no more than 1,000 rows.

startTableDataImportJob_dataFormat :: Lens' StartTableDataImportJob ImportSourceDataFormat

The format of the data that is being imported. Currently the only option supported is "DELIMITED_TEXT".

startTableDataImportJob_destinationTableId :: Lens' StartTableDataImportJob Text

The ID of the table where the rows are being imported.

If a table with the specified id could not be found, this API throws ResourceNotFoundException.

startTableDataImportJob_importOptions :: Lens' StartTableDataImportJob ImportOptions

The options for customizing this import request.

startTableDataImportJob_clientRequestToken :: Lens' StartTableDataImportJob Text

The request token for performing the update action. Request tokens help to identify duplicate requests. If a call times out or fails due to a transient error like a failed network connection, you can retry the call with the same request token. The service ensures that if the first call using that request token is successfully performed, the second call will not perform the action again.

Note that request tokens are valid only for a few minutes. You cannot use request tokens to dedupe requests spanning hours or days.

Destructuring the Response

data StartTableDataImportJobResponse

See: newStartTableDataImportJobResponse smart constructor.

Constructors

StartTableDataImportJobResponse'

Fields

httpStatus :: Int

The response's http status code.

jobId :: Text

The id that is assigned to this import job. Future requests to find out the status of this import job need to send this id in the appropriate parameter in the request.

jobStatus :: TableDataImportJobStatus

The status of the import job immediately after submitting the request.

newStartTableDataImportJobResponse

Arguments

:: Int
$sel:httpStatus:StartTableDataImportJobResponse'

-> Text
$sel:jobId:StartTableDataImportJobResponse'

-> TableDataImportJobStatus
$sel:jobStatus:StartTableDataImportJobResponse'

-> StartTableDataImportJobResponse

Create a value of StartTableDataImportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StartTableDataImportJobResponse', startTableDataImportJobResponse_httpStatus - The response's http status code.

$sel:jobId:StartTableDataImportJobResponse', startTableDataImportJobResponse_jobId - The id that is assigned to this import job. Future requests to find out the status of this import job need to send this id in the appropriate parameter in the request.

$sel:jobStatus:StartTableDataImportJobResponse', startTableDataImportJobResponse_jobStatus - The status of the import job immediately after submitting the request.

Response Lenses

startTableDataImportJobResponse_httpStatus :: Lens' StartTableDataImportJobResponse Int

The response's http status code.

startTableDataImportJobResponse_jobId :: Lens' StartTableDataImportJobResponse Text

The id that is assigned to this import job. Future requests to find out the status of this import job need to send this id in the appropriate parameter in the request.

startTableDataImportJobResponse_jobStatus :: Lens' StartTableDataImportJobResponse TableDataImportJobStatus

The status of the import job immediately after submitting the request.