Network.AWS.HoneyCode.Types.ImportJobSubmitter

Description

Documentation

data ImportJobSubmitter

An object that contains the attributes of the submitter of the import job.

See: newImportJobSubmitter smart constructor.

Constructors

ImportJobSubmitter'

Fields

email :: Maybe (Sensitive Text)

The email id of the submitter of the import job, if available.

userArn :: Maybe Text

The AWS user ARN of the submitter of the import job, if available.

newImportJobSubmitter :: ImportJobSubmitter

Create a value of ImportJobSubmitter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:email:ImportJobSubmitter', importJobSubmitter_email - The email id of the submitter of the import job, if available.

$sel:userArn:ImportJobSubmitter', importJobSubmitter_userArn - The AWS user ARN of the submitter of the import job, if available.

importJobSubmitter_email :: Lens' ImportJobSubmitter (Maybe Text)

The email id of the submitter of the import job, if available.

importJobSubmitter_userArn :: Lens' ImportJobSubmitter (Maybe Text)

The AWS user ARN of the submitter of the import job, if available.