Network.AWS.IAM.AttachGroupPolicy

Description

Attaches the specified managed policy to the specified IAM group.

You use this operation to attach a managed policy to a group. To embed an inline policy in a group, use PutGroupPolicy.

As a best practice, you can validate your IAM policies. To learn more, see Validating IAM policies in the IAM User Guide.

For more information about policies, see Managed policies and inline policies in the IAM User Guide.

Creating a Request

data AttachGroupPolicy

See: newAttachGroupPolicy smart constructor.

Constructors

AttachGroupPolicy'

Fields

groupName :: Text

The name (friendly name, not ARN) of the group to attach the policy to.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

policyArn :: Text

The Amazon Resource Name (ARN) of the IAM policy you want to attach.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

newAttachGroupPolicy

Arguments

:: Text
$sel:groupName:AttachGroupPolicy'

-> Text
$sel:policyArn:AttachGroupPolicy'

-> AttachGroupPolicy

Create a value of AttachGroupPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupName:AttachGroupPolicy', attachGroupPolicy_groupName - The name (friendly name, not ARN) of the group to attach the policy to.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:policyArn:AttachGroupPolicy', attachGroupPolicy_policyArn - The Amazon Resource Name (ARN) of the IAM policy you want to attach.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

Request Lenses

attachGroupPolicy_groupName :: Lens' AttachGroupPolicy Text

The name (friendly name, not ARN) of the group to attach the policy to.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

attachGroupPolicy_policyArn :: Lens' AttachGroupPolicy Text

The Amazon Resource Name (ARN) of the IAM policy you want to attach.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

Destructuring the Response

data AttachGroupPolicyResponse

See: newAttachGroupPolicyResponse smart constructor.

Constructors

AttachGroupPolicyResponse'

newAttachGroupPolicyResponse :: AttachGroupPolicyResponse

Create a value of AttachGroupPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.