Network.AWS.IAM.CreateLoginProfile

Description

Creates a password for the specified IAM user. A password allows an IAM user to access Amazon Web Services services through the Amazon Web Services Management Console.

You can use the CLI, the Amazon Web Services API, or the Users page in the IAM console to create a password for any IAM user. Use ChangePassword to update your own existing password in the __My Security Credentials__ page in the Amazon Web Services Management Console.

For more information about managing passwords, see Managing passwords in the IAM User Guide.

Creating a Request

data CreateLoginProfile

See: newCreateLoginProfile smart constructor.

Constructors

CreateLoginProfile'

Fields

passwordResetRequired :: Maybe Bool

Specifies whether the user is required to set a new password on next sign-in.

userName :: Text

The name of the IAM user to create a password for. The user must already exist.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

password :: Sensitive Text

The new password for the user.

The regex pattern that is used to validate this parameter is a string of characters. That string can include almost any printable ASCII character from the space (\u0020) through the end of the ASCII character range (\u00FF). You can also include the tab (\u0009), line feed (\u000A), and carriage return (\u000D) characters. Any of these characters are valid in a password. However, many tools, such as the Amazon Web Services Management Console, might restrict the ability to type certain characters because they have special meaning within that tool.

newCreateLoginProfile

Arguments

:: Text
$sel:userName:CreateLoginProfile'

-> Text
$sel:password:CreateLoginProfile'

-> CreateLoginProfile

Create a value of CreateLoginProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:passwordResetRequired:CreateLoginProfile', createLoginProfile_passwordResetRequired - Specifies whether the user is required to set a new password on next sign-in.

$sel:userName:CreateLoginProfile', createLoginProfile_userName - The name of the IAM user to create a password for. The user must already exist.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:password:CreateLoginProfile', createLoginProfile_password - The new password for the user.

The regex pattern that is used to validate this parameter is a string of characters. That string can include almost any printable ASCII character from the space (\u0020) through the end of the ASCII character range (\u00FF). You can also include the tab (\u0009), line feed (\u000A), and carriage return (\u000D) characters. Any of these characters are valid in a password. However, many tools, such as the Amazon Web Services Management Console, might restrict the ability to type certain characters because they have special meaning within that tool.

Request Lenses

createLoginProfile_passwordResetRequired :: Lens' CreateLoginProfile (Maybe Bool)

Specifies whether the user is required to set a new password on next sign-in.

createLoginProfile_userName :: Lens' CreateLoginProfile Text

The name of the IAM user to create a password for. The user must already exist.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

createLoginProfile_password :: Lens' CreateLoginProfile Text

The new password for the user.

The regex pattern that is used to validate this parameter is a string of characters. That string can include almost any printable ASCII character from the space (\u0020) through the end of the ASCII character range (\u00FF). You can also include the tab (\u0009), line feed (\u000A), and carriage return (\u000D) characters. Any of these characters are valid in a password. However, many tools, such as the Amazon Web Services Management Console, might restrict the ability to type certain characters because they have special meaning within that tool.

Destructuring the Response

data CreateLoginProfileResponse

Contains the response to a successful CreateLoginProfile request.

See: newCreateLoginProfileResponse smart constructor.

Constructors

CreateLoginProfileResponse'

Fields

httpStatus :: Int

The response's http status code.

loginProfile :: LoginProfile

A structure containing the user name and password create date.

newCreateLoginProfileResponse

Arguments

:: Int
$sel:httpStatus:CreateLoginProfileResponse'

-> LoginProfile
$sel:loginProfile:CreateLoginProfileResponse'

-> CreateLoginProfileResponse

Create a value of CreateLoginProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateLoginProfileResponse', createLoginProfileResponse_httpStatus - The response's http status code.

$sel:loginProfile:CreateLoginProfileResponse', createLoginProfileResponse_loginProfile - A structure containing the user name and password create date.

Response Lenses

createLoginProfileResponse_httpStatus :: Lens' CreateLoginProfileResponse Int

The response's http status code.

createLoginProfileResponse_loginProfile :: Lens' CreateLoginProfileResponse LoginProfile

A structure containing the user name and password create date.