Network.AWS.IAM.CreateSAMLProvider

Description

Creates an IAM resource that describes an identity provider (IdP) that supports SAML 2.0.

The SAML provider resource that you create with this operation can be used as a principal in an IAM role's trust policy. Such a policy can enable federated users who sign in using the SAML IdP to assume the role. You can create an IAM role that supports Web-based single sign-on (SSO) to the Amazon Web Services Management Console or one that supports API access to Amazon Web Services.

When you create the SAML provider resource, you upload a SAML metadata document that you get from your IdP. That document includes the issuer's name, expiration information, and keys that can be used to validate the SAML authentication response (assertions) that the IdP sends. You must generate the metadata document using the identity management software that is used as your organization's IdP.

This operation requires Signature Version 4.

For more information, see Enabling SAML 2.0 federated users to access the Amazon Web Services Management Console and About SAML 2.0-based federation in the IAM User Guide.

Creating a Request

data CreateSAMLProvider

See: newCreateSAMLProvider smart constructor.

Constructors

CreateSAMLProvider'

Fields

tags :: Maybe [Tag]

A list of tags that you want to attach to the new IAM SAML provider. Each tag consists of a key name and an associated value. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

If any one of the tags is invalid or if you exceed the allowed maximum number of tags, then the entire request fails and the resource is not created.

sAMLMetadataDocument :: Text

An XML document generated by an identity provider (IdP) that supports SAML 2.0. The document includes the issuer's name, expiration information, and keys that can be used to validate the SAML authentication response (assertions) that are received from the IdP. You must generate the metadata document using the identity management software that is used as your organization's IdP.

For more information, see About SAML 2.0-based federation in the IAM User Guide

name :: Text

The name of the provider to create.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

newCreateSAMLProvider

Arguments

:: Text
$sel:sAMLMetadataDocument:CreateSAMLProvider'

-> Text
$sel:name:CreateSAMLProvider'

-> CreateSAMLProvider

Create a value of CreateSAMLProvider with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateSAMLProvider', createSAMLProvider_tags - A list of tags that you want to attach to the new IAM SAML provider. Each tag consists of a key name and an associated value. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

If any one of the tags is invalid or if you exceed the allowed maximum number of tags, then the entire request fails and the resource is not created.

$sel:sAMLMetadataDocument:CreateSAMLProvider', createSAMLProvider_sAMLMetadataDocument - An XML document generated by an identity provider (IdP) that supports SAML 2.0. The document includes the issuer's name, expiration information, and keys that can be used to validate the SAML authentication response (assertions) that are received from the IdP. You must generate the metadata document using the identity management software that is used as your organization's IdP.

For more information, see About SAML 2.0-based federation in the IAM User Guide

$sel:name:CreateSAMLProvider', createSAMLProvider_name - The name of the provider to create.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Request Lenses

createSAMLProvider_tags :: Lens' CreateSAMLProvider (Maybe [Tag])

A list of tags that you want to attach to the new IAM SAML provider. Each tag consists of a key name and an associated value. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

If any one of the tags is invalid or if you exceed the allowed maximum number of tags, then the entire request fails and the resource is not created.

createSAMLProvider_sAMLMetadataDocument :: Lens' CreateSAMLProvider Text

An XML document generated by an identity provider (IdP) that supports SAML 2.0. The document includes the issuer's name, expiration information, and keys that can be used to validate the SAML authentication response (assertions) that are received from the IdP. You must generate the metadata document using the identity management software that is used as your organization's IdP.

For more information, see About SAML 2.0-based federation in the IAM User Guide

createSAMLProvider_name :: Lens' CreateSAMLProvider Text

The name of the provider to create.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Destructuring the Response

data CreateSAMLProviderResponse

Contains the response to a successful CreateSAMLProvider request.

See: newCreateSAMLProviderResponse smart constructor.

Constructors

CreateSAMLProviderResponse'

Fields

sAMLProviderArn :: Maybe Text

The Amazon Resource Name (ARN) of the new SAML provider resource in IAM.

tags :: Maybe [Tag]

A list of tags that are attached to the new IAM SAML provider. The returned list of tags is sorted by tag key. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

httpStatus :: Int

The response's http status code.

newCreateSAMLProviderResponse

Arguments

:: Int
$sel:httpStatus:CreateSAMLProviderResponse'

-> CreateSAMLProviderResponse

Create a value of CreateSAMLProviderResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sAMLProviderArn:CreateSAMLProviderResponse', createSAMLProviderResponse_sAMLProviderArn - The Amazon Resource Name (ARN) of the new SAML provider resource in IAM.

$sel:tags:CreateSAMLProvider', createSAMLProviderResponse_tags - A list of tags that are attached to the new IAM SAML provider. The returned list of tags is sorted by tag key. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

$sel:httpStatus:CreateSAMLProviderResponse', createSAMLProviderResponse_httpStatus - The response's http status code.

Response Lenses

createSAMLProviderResponse_sAMLProviderArn :: Lens' CreateSAMLProviderResponse (Maybe Text)

The Amazon Resource Name (ARN) of the new SAML provider resource in IAM.

createSAMLProviderResponse_tags :: Lens' CreateSAMLProviderResponse (Maybe [Tag])

A list of tags that are attached to the new IAM SAML provider. The returned list of tags is sorted by tag key. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

createSAMLProviderResponse_httpStatus :: Lens' CreateSAMLProviderResponse Int

The response's http status code.