Network.AWS.IAM.CreateUser

Description

Creates a new IAM user for your Amazon Web Services account.

For information about quotas for the number of IAM users you can create, see IAM and STS quotas in the IAM User Guide.

Creating a Request

data CreateUser

See: newCreateUser smart constructor.

Constructors

CreateUser'

Fields

permissionsBoundary :: Maybe Text

The ARN of the policy that is used to set the permissions boundary for the user.

tags :: Maybe [Tag]

A list of tags that you want to attach to the new user. Each tag consists of a key name and an associated value. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

If any one of the tags is invalid or if you exceed the allowed maximum number of tags, then the entire request fails and the resource is not created.

path :: Maybe Text

The path for the user name. For more information about paths, see IAM identifiers in the IAM User Guide.

This parameter is optional. If it is not included, it defaults to a slash (/).

This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

userName :: Text

The name of the user to create.

IAM user, group, role, and policy names must be unique within the account. Names are not distinguished by case. For example, you cannot create resources named both "MyResource" and "myresource".

newCreateUser

Arguments

:: Text
$sel:userName:CreateUser'

-> CreateUser

Create a value of CreateUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:permissionsBoundary:CreateUser', createUser_permissionsBoundary - The ARN of the policy that is used to set the permissions boundary for the user.

$sel:tags:CreateUser', createUser_tags - A list of tags that you want to attach to the new user. Each tag consists of a key name and an associated value. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

If any one of the tags is invalid or if you exceed the allowed maximum number of tags, then the entire request fails and the resource is not created.

$sel:path:CreateUser', createUser_path - The path for the user name. For more information about paths, see IAM identifiers in the IAM User Guide.

This parameter is optional. If it is not included, it defaults to a slash (/).

This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

$sel:userName:CreateUser', createUser_userName - The name of the user to create.

IAM user, group, role, and policy names must be unique within the account. Names are not distinguished by case. For example, you cannot create resources named both "MyResource" and "myresource".

Request Lenses

createUser_permissionsBoundary :: Lens' CreateUser (Maybe Text)

The ARN of the policy that is used to set the permissions boundary for the user.

createUser_tags :: Lens' CreateUser (Maybe [Tag])

A list of tags that you want to attach to the new user. Each tag consists of a key name and an associated value. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

If any one of the tags is invalid or if you exceed the allowed maximum number of tags, then the entire request fails and the resource is not created.

createUser_path :: Lens' CreateUser (Maybe Text)

The path for the user name. For more information about paths, see IAM identifiers in the IAM User Guide.

This parameter is optional. If it is not included, it defaults to a slash (/).

This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

createUser_userName :: Lens' CreateUser Text

The name of the user to create.

IAM user, group, role, and policy names must be unique within the account. Names are not distinguished by case. For example, you cannot create resources named both "MyResource" and "myresource".

Destructuring the Response

data CreateUserResponse

Contains the response to a successful CreateUser request.

See: newCreateUserResponse smart constructor.

Constructors

CreateUserResponse'

Fields

user :: Maybe User

A structure with details about the new IAM user.

httpStatus :: Int

The response's http status code.

newCreateUserResponse

Arguments

:: Int
$sel:httpStatus:CreateUserResponse'

-> CreateUserResponse

Create a value of CreateUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:user:CreateUserResponse', createUserResponse_user - A structure with details about the new IAM user.

$sel:httpStatus:CreateUserResponse', createUserResponse_httpStatus - The response's http status code.

Response Lenses

createUserResponse_user :: Lens' CreateUserResponse (Maybe User)

A structure with details about the new IAM user.

createUserResponse_httpStatus :: Lens' CreateUserResponse Int

The response's http status code.