Network.AWS.IAM.DeactivateMFADevice

Description

Deactivates the specified MFA device and removes it from association with the user name for which it was originally enabled.

For more information about creating and working with virtual MFA devices, see Enabling a virtual multi-factor authentication (MFA) device in the IAM User Guide.

Creating a Request

data DeactivateMFADevice

See: newDeactivateMFADevice smart constructor.

Constructors

DeactivateMFADevice'

Fields

userName :: Text

The name of the user whose MFA device you want to deactivate.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

serialNumber :: Text

The serial number that uniquely identifies the MFA device. For virtual MFA devices, the serial number is the device ARN.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@:/-

newDeactivateMFADevice

Arguments

:: Text
$sel:userName:DeactivateMFADevice'

-> Text
$sel:serialNumber:DeactivateMFADevice'

-> DeactivateMFADevice

Create a value of DeactivateMFADevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userName:DeactivateMFADevice', deactivateMFADevice_userName - The name of the user whose MFA device you want to deactivate.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:serialNumber:DeactivateMFADevice', deactivateMFADevice_serialNumber - The serial number that uniquely identifies the MFA device. For virtual MFA devices, the serial number is the device ARN.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@:/-

Request Lenses

deactivateMFADevice_userName :: Lens' DeactivateMFADevice Text

The name of the user whose MFA device you want to deactivate.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

deactivateMFADevice_serialNumber :: Lens' DeactivateMFADevice Text

The serial number that uniquely identifies the MFA device. For virtual MFA devices, the serial number is the device ARN.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: =,.@:/-

Destructuring the Response

data DeactivateMFADeviceResponse

See: newDeactivateMFADeviceResponse smart constructor.

Constructors

DeactivateMFADeviceResponse'

newDeactivateMFADeviceResponse :: DeactivateMFADeviceResponse

Create a value of DeactivateMFADeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.