Network.AWS.IAM.DeleteAccessKey

Description

Deletes the access key pair associated with the specified IAM user.

If you do not specify a user name, IAM determines the user name implicitly based on the Amazon Web Services access key ID signing the request. This operation works for access keys under the Amazon Web Services account. Consequently, you can use this operation to manage Amazon Web Services account root user credentials even if the Amazon Web Services account has no associated users.

Creating a Request

data DeleteAccessKey

See: newDeleteAccessKey smart constructor.

Constructors

DeleteAccessKey'

Fields

userName :: Maybe Text

The name of the user whose access key pair you want to delete.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

accessKeyId :: AccessKey

The access key ID for the access key ID and secret access key you want to delete.

This parameter allows (through its regex pattern) a string of characters that can consist of any upper or lowercased letter or digit.

newDeleteAccessKey

Arguments

:: AccessKey
$sel:accessKeyId:DeleteAccessKey'

-> DeleteAccessKey

Create a value of DeleteAccessKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userName:DeleteAccessKey', deleteAccessKey_userName - The name of the user whose access key pair you want to delete.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:accessKeyId:DeleteAccessKey', deleteAccessKey_accessKeyId - The access key ID for the access key ID and secret access key you want to delete.

This parameter allows (through its regex pattern) a string of characters that can consist of any upper or lowercased letter or digit.

Request Lenses

deleteAccessKey_userName :: Lens' DeleteAccessKey (Maybe Text)

The name of the user whose access key pair you want to delete.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

deleteAccessKey_accessKeyId :: Lens' DeleteAccessKey AccessKey

The access key ID for the access key ID and secret access key you want to delete.

This parameter allows (through its regex pattern) a string of characters that can consist of any upper or lowercased letter or digit.

Destructuring the Response

data DeleteAccessKeyResponse

See: newDeleteAccessKeyResponse smart constructor.

Constructors

DeleteAccessKeyResponse'

newDeleteAccessKeyResponse :: DeleteAccessKeyResponse

Create a value of DeleteAccessKeyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.