Network.AWS.IAM.DeleteGroup

Description

Deletes the specified IAM group. The group must not contain any users or have any attached policies.

Creating a Request

data DeleteGroup

See: newDeleteGroup smart constructor.

Constructors

DeleteGroup'

Fields

groupName :: Text

The name of the IAM group to delete.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

newDeleteGroup

Arguments

:: Text
$sel:groupName:DeleteGroup'

-> DeleteGroup

Create a value of DeleteGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupName:DeleteGroup', deleteGroup_groupName - The name of the IAM group to delete.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Request Lenses

deleteGroup_groupName :: Lens' DeleteGroup Text

The name of the IAM group to delete.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Destructuring the Response

data DeleteGroupResponse

See: newDeleteGroupResponse smart constructor.

Constructors

DeleteGroupResponse'

newDeleteGroupResponse :: DeleteGroupResponse

Create a value of DeleteGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.