Network.AWS.IAM.DeleteSSHPublicKey

Description

Deletes the specified SSH public key.

The SSH public key deleted by this operation is used only for authenticating the associated IAM user to an CodeCommit repository. For more information about using SSH keys to authenticate to an CodeCommit repository, see Set up CodeCommit for SSH connections in the CodeCommit User Guide.

Creating a Request

data DeleteSSHPublicKey

See: newDeleteSSHPublicKey smart constructor.

Constructors

DeleteSSHPublicKey'

Fields

userName :: Text

The name of the IAM user associated with the SSH public key.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

sSHPublicKeyId :: Text

The unique identifier for the SSH public key.

This parameter allows (through its regex pattern) a string of characters that can consist of any upper or lowercased letter or digit.

newDeleteSSHPublicKey

Arguments

:: Text
$sel:userName:DeleteSSHPublicKey'

-> Text
$sel:sSHPublicKeyId:DeleteSSHPublicKey'

-> DeleteSSHPublicKey

Create a value of DeleteSSHPublicKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userName:DeleteSSHPublicKey', deleteSSHPublicKey_userName - The name of the IAM user associated with the SSH public key.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:sSHPublicKeyId:DeleteSSHPublicKey', deleteSSHPublicKey_sSHPublicKeyId - The unique identifier for the SSH public key.

This parameter allows (through its regex pattern) a string of characters that can consist of any upper or lowercased letter or digit.

Request Lenses

deleteSSHPublicKey_userName :: Lens' DeleteSSHPublicKey Text

The name of the IAM user associated with the SSH public key.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

deleteSSHPublicKey_sSHPublicKeyId :: Lens' DeleteSSHPublicKey Text

The unique identifier for the SSH public key.

This parameter allows (through its regex pattern) a string of characters that can consist of any upper or lowercased letter or digit.

Destructuring the Response

data DeleteSSHPublicKeyResponse

See: newDeleteSSHPublicKeyResponse smart constructor.

Constructors

DeleteSSHPublicKeyResponse'

newDeleteSSHPublicKeyResponse :: DeleteSSHPublicKeyResponse

Create a value of DeleteSSHPublicKeyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.