Network.AWS.IAM.DetachUserPolicy

Description

Removes the specified managed policy from the specified user.

A user can also have inline policies embedded with it. To delete an inline policy, use DeleteUserPolicy. For information about policies, see Managed policies and inline policies in the IAM User Guide.

Creating a Request

data DetachUserPolicy

See: newDetachUserPolicy smart constructor.

Constructors

DetachUserPolicy'

Fields

userName :: Text

The name (friendly name, not ARN) of the IAM user to detach the policy from.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

policyArn :: Text

The Amazon Resource Name (ARN) of the IAM policy you want to detach.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

newDetachUserPolicy

Arguments

:: Text
$sel:userName:DetachUserPolicy'

-> Text
$sel:policyArn:DetachUserPolicy'

-> DetachUserPolicy

Create a value of DetachUserPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userName:DetachUserPolicy', detachUserPolicy_userName - The name (friendly name, not ARN) of the IAM user to detach the policy from.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:policyArn:DetachUserPolicy', detachUserPolicy_policyArn - The Amazon Resource Name (ARN) of the IAM policy you want to detach.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

Request Lenses

detachUserPolicy_userName :: Lens' DetachUserPolicy Text

The name (friendly name, not ARN) of the IAM user to detach the policy from.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

detachUserPolicy_policyArn :: Lens' DetachUserPolicy Text

The Amazon Resource Name (ARN) of the IAM policy you want to detach.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

Destructuring the Response

data DetachUserPolicyResponse

See: newDetachUserPolicyResponse smart constructor.

Constructors

DetachUserPolicyResponse'

newDetachUserPolicyResponse :: DetachUserPolicyResponse

Create a value of DetachUserPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.