Network.AWS.IAM.GetAccessKeyLastUsed

Description

Retrieves information about when the specified access key was last used. The information includes the date and time of last use, along with the Amazon Web Services service and Region that were specified in the last request made with that key.

Creating a Request

data GetAccessKeyLastUsed

See: newGetAccessKeyLastUsed smart constructor.

Constructors

GetAccessKeyLastUsed'

Fields

accessKeyId :: AccessKey

The identifier of an access key.

This parameter allows (through its regex pattern) a string of characters that can consist of any upper or lowercased letter or digit.

newGetAccessKeyLastUsed

Arguments

:: AccessKey
$sel:accessKeyId:GetAccessKeyLastUsed'

-> GetAccessKeyLastUsed

Create a value of GetAccessKeyLastUsed with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessKeyId:GetAccessKeyLastUsed', getAccessKeyLastUsed_accessKeyId - The identifier of an access key.

This parameter allows (through its regex pattern) a string of characters that can consist of any upper or lowercased letter or digit.

Request Lenses

getAccessKeyLastUsed_accessKeyId :: Lens' GetAccessKeyLastUsed AccessKey

The identifier of an access key.

This parameter allows (through its regex pattern) a string of characters that can consist of any upper or lowercased letter or digit.

Destructuring the Response

data GetAccessKeyLastUsedResponse

Contains the response to a successful GetAccessKeyLastUsed request. It is also returned as a member of the AccessKeyMetaData structure returned by the ListAccessKeys action.

See: newGetAccessKeyLastUsedResponse smart constructor.

Constructors

GetAccessKeyLastUsedResponse'

Fields

userName :: Maybe Text

The name of the IAM user that owns this access key.

accessKeyLastUsed :: Maybe AccessKeyLastUsed

Contains information about the last time the access key was used.

httpStatus :: Int

The response's http status code.

newGetAccessKeyLastUsedResponse

Arguments

:: Int
$sel:httpStatus:GetAccessKeyLastUsedResponse'

-> GetAccessKeyLastUsedResponse

Create a value of GetAccessKeyLastUsedResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userName:GetAccessKeyLastUsedResponse', getAccessKeyLastUsedResponse_userName - The name of the IAM user that owns this access key.

$sel:accessKeyLastUsed:GetAccessKeyLastUsedResponse', getAccessKeyLastUsedResponse_accessKeyLastUsed - Contains information about the last time the access key was used.

$sel:httpStatus:GetAccessKeyLastUsedResponse', getAccessKeyLastUsedResponse_httpStatus - The response's http status code.

Response Lenses

getAccessKeyLastUsedResponse_userName :: Lens' GetAccessKeyLastUsedResponse (Maybe Text)

The name of the IAM user that owns this access key.

getAccessKeyLastUsedResponse_accessKeyLastUsed :: Lens' GetAccessKeyLastUsedResponse (Maybe AccessKeyLastUsed)

Contains information about the last time the access key was used.

getAccessKeyLastUsedResponse_httpStatus :: Lens' GetAccessKeyLastUsedResponse Int

The response's http status code.