Network.AWS.IAM.GetAccountAuthorizationDetails

Description

Retrieves information about all IAM users, groups, roles, and policies in your Amazon Web Services account, including their relationships to one another. Use this operation to obtain a snapshot of the configuration of IAM permissions (users, groups, roles, and policies) in your account.

Policies returned by this operation are URL-encoded compliant with RFC 3986. You can use a URL decoding method to convert the policy back to plain JSON text. For example, if you use Java, you can use the decode method of the java.net.URLDecoder utility class in the Java SDK. Other languages and SDKs provide similar functionality.

You can optionally filter the results using the Filter parameter. You can paginate the results using the MaxItems and Marker parameters.

This operation returns paginated results.

Creating a Request

data GetAccountAuthorizationDetails

See: newGetAccountAuthorizationDetails smart constructor.

Constructors

GetAccountAuthorizationDetails'

Fields

filter' :: Maybe [EntityType]

A list of entity types used to filter the results. Only the entities that match the types you specify are included in the output. Use the value LocalManagedPolicy to include customer managed policies.

The format for this parameter is a comma-separated (if more than one) list of strings. Each string value in the list must be one of the valid values listed below.

maxItems :: Maybe Natural

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

marker :: Maybe Text

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

newGetAccountAuthorizationDetails :: GetAccountAuthorizationDetails

Create a value of GetAccountAuthorizationDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filter':GetAccountAuthorizationDetails', getAccountAuthorizationDetails_filter - A list of entity types used to filter the results. Only the entities that match the types you specify are included in the output. Use the value LocalManagedPolicy to include customer managed policies.

The format for this parameter is a comma-separated (if more than one) list of strings. Each string value in the list must be one of the valid values listed below.

$sel:maxItems:GetAccountAuthorizationDetails', getAccountAuthorizationDetails_maxItems - Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

$sel:marker:GetAccountAuthorizationDetails', getAccountAuthorizationDetails_marker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

Request Lenses

getAccountAuthorizationDetails_filter :: Lens' GetAccountAuthorizationDetails (Maybe [EntityType])

A list of entity types used to filter the results. Only the entities that match the types you specify are included in the output. Use the value LocalManagedPolicy to include customer managed policies.

The format for this parameter is a comma-separated (if more than one) list of strings. Each string value in the list must be one of the valid values listed below.

getAccountAuthorizationDetails_maxItems :: Lens' GetAccountAuthorizationDetails (Maybe Natural)

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

getAccountAuthorizationDetails_marker :: Lens' GetAccountAuthorizationDetails (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

Destructuring the Response

data GetAccountAuthorizationDetailsResponse

Contains the response to a successful GetAccountAuthorizationDetails request.

See: newGetAccountAuthorizationDetailsResponse smart constructor.

Constructors

GetAccountAuthorizationDetailsResponse'

Fields

roleDetailList :: Maybe [RoleDetail]

A list containing information about IAM roles.

groupDetailList :: Maybe [GroupDetail]

A list containing information about IAM groups.

policies :: Maybe [ManagedPolicyDetail]

A list containing information about managed policies.

isTruncated :: Maybe Bool

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

userDetailList :: Maybe [UserDetail]

A list containing information about IAM users.

marker :: Maybe Text

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

httpStatus :: Int

The response's http status code.

newGetAccountAuthorizationDetailsResponse

Arguments

:: Int
$sel:httpStatus:GetAccountAuthorizationDetailsResponse'

-> GetAccountAuthorizationDetailsResponse

Create a value of GetAccountAuthorizationDetailsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleDetailList:GetAccountAuthorizationDetailsResponse', getAccountAuthorizationDetailsResponse_roleDetailList - A list containing information about IAM roles.

$sel:groupDetailList:GetAccountAuthorizationDetailsResponse', getAccountAuthorizationDetailsResponse_groupDetailList - A list containing information about IAM groups.

$sel:policies:GetAccountAuthorizationDetailsResponse', getAccountAuthorizationDetailsResponse_policies - A list containing information about managed policies.

$sel:isTruncated:GetAccountAuthorizationDetailsResponse', getAccountAuthorizationDetailsResponse_isTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

$sel:userDetailList:GetAccountAuthorizationDetailsResponse', getAccountAuthorizationDetailsResponse_userDetailList - A list containing information about IAM users.

$sel:marker:GetAccountAuthorizationDetails', getAccountAuthorizationDetailsResponse_marker - When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

$sel:httpStatus:GetAccountAuthorizationDetailsResponse', getAccountAuthorizationDetailsResponse_httpStatus - The response's http status code.

Response Lenses

getAccountAuthorizationDetailsResponse_roleDetailList :: Lens' GetAccountAuthorizationDetailsResponse (Maybe [RoleDetail])

A list containing information about IAM roles.

getAccountAuthorizationDetailsResponse_groupDetailList :: Lens' GetAccountAuthorizationDetailsResponse (Maybe [GroupDetail])

A list containing information about IAM groups.

getAccountAuthorizationDetailsResponse_policies :: Lens' GetAccountAuthorizationDetailsResponse (Maybe [ManagedPolicyDetail])

A list containing information about managed policies.

getAccountAuthorizationDetailsResponse_isTruncated :: Lens' GetAccountAuthorizationDetailsResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

getAccountAuthorizationDetailsResponse_userDetailList :: Lens' GetAccountAuthorizationDetailsResponse (Maybe [UserDetail])

A list containing information about IAM users.

getAccountAuthorizationDetailsResponse_marker :: Lens' GetAccountAuthorizationDetailsResponse (Maybe Text)

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

getAccountAuthorizationDetailsResponse_httpStatus :: Lens' GetAccountAuthorizationDetailsResponse Int

The response's http status code.