Network.AWS.IAM.GetOpenIDConnectProvider

Description

Returns information about the specified OpenID Connect (OIDC) provider resource object in IAM.

Creating a Request

data GetOpenIDConnectProvider

See: newGetOpenIDConnectProvider smart constructor.

Constructors

GetOpenIDConnectProvider'

Fields

openIDConnectProviderArn :: Text

The Amazon Resource Name (ARN) of the OIDC provider resource object in IAM to get information for. You can get a list of OIDC provider resource ARNs by using the ListOpenIDConnectProviders operation.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

newGetOpenIDConnectProvider

Arguments

:: Text
$sel:openIDConnectProviderArn:GetOpenIDConnectProvider'

-> GetOpenIDConnectProvider

Create a value of GetOpenIDConnectProvider with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:openIDConnectProviderArn:GetOpenIDConnectProvider', getOpenIDConnectProvider_openIDConnectProviderArn - The Amazon Resource Name (ARN) of the OIDC provider resource object in IAM to get information for. You can get a list of OIDC provider resource ARNs by using the ListOpenIDConnectProviders operation.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

Request Lenses

getOpenIDConnectProvider_openIDConnectProviderArn :: Lens' GetOpenIDConnectProvider Text

The Amazon Resource Name (ARN) of the OIDC provider resource object in IAM to get information for. You can get a list of OIDC provider resource ARNs by using the ListOpenIDConnectProviders operation.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

Destructuring the Response

data GetOpenIDConnectProviderResponse

Contains the response to a successful GetOpenIDConnectProvider request.

See: newGetOpenIDConnectProviderResponse smart constructor.

Constructors

GetOpenIDConnectProviderResponse'

Fields

clientIDList :: Maybe [Text]

A list of client IDs (also known as audiences) that are associated with the specified IAM OIDC provider resource object. For more information, see CreateOpenIDConnectProvider.

createDate :: Maybe ISO8601

The date and time when the IAM OIDC provider resource object was created in the Amazon Web Services account.

thumbprintList :: Maybe [Text]

A list of certificate thumbprints that are associated with the specified IAM OIDC provider resource object. For more information, see CreateOpenIDConnectProvider.

tags :: Maybe [Tag]

A list of tags that are attached to the specified IAM OIDC provider. The returned list of tags is sorted by tag key. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

url :: Maybe Text

The URL that the IAM OIDC provider resource object is associated with. For more information, see CreateOpenIDConnectProvider.

httpStatus :: Int

The response's http status code.

newGetOpenIDConnectProviderResponse

Arguments

:: Int
$sel:httpStatus:GetOpenIDConnectProviderResponse'

-> GetOpenIDConnectProviderResponse

Create a value of GetOpenIDConnectProviderResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientIDList:GetOpenIDConnectProviderResponse', getOpenIDConnectProviderResponse_clientIDList - A list of client IDs (also known as audiences) that are associated with the specified IAM OIDC provider resource object. For more information, see CreateOpenIDConnectProvider.

$sel:createDate:GetOpenIDConnectProviderResponse', getOpenIDConnectProviderResponse_createDate - The date and time when the IAM OIDC provider resource object was created in the Amazon Web Services account.

$sel:thumbprintList:GetOpenIDConnectProviderResponse', getOpenIDConnectProviderResponse_thumbprintList - A list of certificate thumbprints that are associated with the specified IAM OIDC provider resource object. For more information, see CreateOpenIDConnectProvider.

$sel:tags:GetOpenIDConnectProviderResponse', getOpenIDConnectProviderResponse_tags - A list of tags that are attached to the specified IAM OIDC provider. The returned list of tags is sorted by tag key. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

$sel:url:GetOpenIDConnectProviderResponse', getOpenIDConnectProviderResponse_url - The URL that the IAM OIDC provider resource object is associated with. For more information, see CreateOpenIDConnectProvider.

$sel:httpStatus:GetOpenIDConnectProviderResponse', getOpenIDConnectProviderResponse_httpStatus - The response's http status code.

Response Lenses

getOpenIDConnectProviderResponse_clientIDList :: Lens' GetOpenIDConnectProviderResponse (Maybe [Text])

A list of client IDs (also known as audiences) that are associated with the specified IAM OIDC provider resource object. For more information, see CreateOpenIDConnectProvider.

getOpenIDConnectProviderResponse_createDate :: Lens' GetOpenIDConnectProviderResponse (Maybe UTCTime)

The date and time when the IAM OIDC provider resource object was created in the Amazon Web Services account.

getOpenIDConnectProviderResponse_thumbprintList :: Lens' GetOpenIDConnectProviderResponse (Maybe [Text])

A list of certificate thumbprints that are associated with the specified IAM OIDC provider resource object. For more information, see CreateOpenIDConnectProvider.

getOpenIDConnectProviderResponse_tags :: Lens' GetOpenIDConnectProviderResponse (Maybe [Tag])

A list of tags that are attached to the specified IAM OIDC provider. The returned list of tags is sorted by tag key. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

getOpenIDConnectProviderResponse_url :: Lens' GetOpenIDConnectProviderResponse (Maybe Text)

The URL that the IAM OIDC provider resource object is associated with. For more information, see CreateOpenIDConnectProvider.

getOpenIDConnectProviderResponse_httpStatus :: Lens' GetOpenIDConnectProviderResponse Int

The response's http status code.