Network.AWS.IAM.GetOrganizationsAccessReport

Description

Retrieves the service last accessed data report for Organizations that was previously generated using the GenerateOrganizationsAccessReport operation. This operation retrieves the status of your report job and the report contents.

Depending on the parameters that you passed when you generated the report, the data returned could include different information. For details, see GenerateOrganizationsAccessReport.

To call this operation, you must be signed in to the management account in your organization. SCPs must be enabled for your organization root. You must have permissions to perform this operation. For more information, see Refining permissions using service last accessed data in the IAM User Guide.

For each service that principals in an account (root users, IAM users, or IAM roles) could access using SCPs, the operation returns details about the most recent access attempt. If there was no attempt, the service is listed without details about the most recent attempt to access the service. If the operation fails, it returns the reason that it failed.

By default, the list is sorted by service namespace.

Creating a Request

data GetOrganizationsAccessReport

See: newGetOrganizationsAccessReport smart constructor.

Constructors

GetOrganizationsAccessReport'

Fields

sortKey :: Maybe SortKeyType

The key that is used to sort the results. If you choose the namespace key, the results are returned in alphabetical order. If you choose the time key, the results are sorted numerically by the date and time.

maxItems :: Maybe Natural

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

marker :: Maybe Text

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

jobId :: Text

The identifier of the request generated by the GenerateOrganizationsAccessReport operation.

newGetOrganizationsAccessReport

Arguments

:: Text
$sel:jobId:GetOrganizationsAccessReport'

-> GetOrganizationsAccessReport

Create a value of GetOrganizationsAccessReport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sortKey:GetOrganizationsAccessReport', getOrganizationsAccessReport_sortKey - The key that is used to sort the results. If you choose the namespace key, the results are returned in alphabetical order. If you choose the time key, the results are sorted numerically by the date and time.

$sel:maxItems:GetOrganizationsAccessReport', getOrganizationsAccessReport_maxItems - Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

$sel:marker:GetOrganizationsAccessReport', getOrganizationsAccessReport_marker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

$sel:jobId:GetOrganizationsAccessReport', getOrganizationsAccessReport_jobId - The identifier of the request generated by the GenerateOrganizationsAccessReport operation.

Request Lenses

getOrganizationsAccessReport_sortKey :: Lens' GetOrganizationsAccessReport (Maybe SortKeyType)

The key that is used to sort the results. If you choose the namespace key, the results are returned in alphabetical order. If you choose the time key, the results are sorted numerically by the date and time.

getOrganizationsAccessReport_maxItems :: Lens' GetOrganizationsAccessReport (Maybe Natural)

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

getOrganizationsAccessReport_marker :: Lens' GetOrganizationsAccessReport (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

getOrganizationsAccessReport_jobId :: Lens' GetOrganizationsAccessReport Text

The identifier of the request generated by the GenerateOrganizationsAccessReport operation.

Destructuring the Response

data GetOrganizationsAccessReportResponse

See: newGetOrganizationsAccessReportResponse smart constructor.

Constructors

GetOrganizationsAccessReportResponse'

Fields

accessDetails :: Maybe [AccessDetail]

An object that contains details about the most recent attempt to access the service.

isTruncated :: Maybe Bool

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

jobCompletionDate :: Maybe ISO8601

The date and time, in ISO 8601 date-time format, when the generated report job was completed or failed.

This field is null if the job is still in progress, as indicated by a job status value of IN_PROGRESS.

numberOfServicesNotAccessed :: Maybe Int

The number of services that account principals are allowed but did not attempt to access.

numberOfServicesAccessible :: Maybe Int

The number of services that the applicable SCPs allow account principals to access.

marker :: Maybe Text

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

errorDetails :: Maybe ErrorDetails

httpStatus :: Int

The response's http status code.

jobStatus :: JobStatusType

The status of the job.

jobCreationDate :: ISO8601

The date and time, in ISO 8601 date-time format, when the report job was created.

newGetOrganizationsAccessReportResponse

Arguments

:: Int
$sel:httpStatus:GetOrganizationsAccessReportResponse'

-> JobStatusType
$sel:jobStatus:GetOrganizationsAccessReportResponse'

-> UTCTime
$sel:jobCreationDate:GetOrganizationsAccessReportResponse'

-> GetOrganizationsAccessReportResponse

Create a value of GetOrganizationsAccessReportResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessDetails:GetOrganizationsAccessReportResponse', getOrganizationsAccessReportResponse_accessDetails - An object that contains details about the most recent attempt to access the service.

$sel:isTruncated:GetOrganizationsAccessReportResponse', getOrganizationsAccessReportResponse_isTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

$sel:jobCompletionDate:GetOrganizationsAccessReportResponse', getOrganizationsAccessReportResponse_jobCompletionDate - The date and time, in ISO 8601 date-time format, when the generated report job was completed or failed.

This field is null if the job is still in progress, as indicated by a job status value of IN_PROGRESS.

$sel:numberOfServicesNotAccessed:GetOrganizationsAccessReportResponse', getOrganizationsAccessReportResponse_numberOfServicesNotAccessed - The number of services that account principals are allowed but did not attempt to access.

$sel:numberOfServicesAccessible:GetOrganizationsAccessReportResponse', getOrganizationsAccessReportResponse_numberOfServicesAccessible - The number of services that the applicable SCPs allow account principals to access.

$sel:marker:GetOrganizationsAccessReport', getOrganizationsAccessReportResponse_marker - When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

$sel:errorDetails:GetOrganizationsAccessReportResponse', getOrganizationsAccessReportResponse_errorDetails - Undocumented member.

$sel:httpStatus:GetOrganizationsAccessReportResponse', getOrganizationsAccessReportResponse_httpStatus - The response's http status code.

$sel:jobStatus:GetOrganizationsAccessReportResponse', getOrganizationsAccessReportResponse_jobStatus - The status of the job.

$sel:jobCreationDate:GetOrganizationsAccessReportResponse', getOrganizationsAccessReportResponse_jobCreationDate - The date and time, in ISO 8601 date-time format, when the report job was created.

Response Lenses

getOrganizationsAccessReportResponse_accessDetails :: Lens' GetOrganizationsAccessReportResponse (Maybe [AccessDetail])

An object that contains details about the most recent attempt to access the service.

getOrganizationsAccessReportResponse_isTruncated :: Lens' GetOrganizationsAccessReportResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

getOrganizationsAccessReportResponse_jobCompletionDate :: Lens' GetOrganizationsAccessReportResponse (Maybe UTCTime)

The date and time, in ISO 8601 date-time format, when the generated report job was completed or failed.

This field is null if the job is still in progress, as indicated by a job status value of IN_PROGRESS.

getOrganizationsAccessReportResponse_numberOfServicesNotAccessed :: Lens' GetOrganizationsAccessReportResponse (Maybe Int)

The number of services that account principals are allowed but did not attempt to access.

getOrganizationsAccessReportResponse_numberOfServicesAccessible :: Lens' GetOrganizationsAccessReportResponse (Maybe Int)

The number of services that the applicable SCPs allow account principals to access.

getOrganizationsAccessReportResponse_marker :: Lens' GetOrganizationsAccessReportResponse (Maybe Text)

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

getOrganizationsAccessReportResponse_errorDetails :: Lens' GetOrganizationsAccessReportResponse (Maybe ErrorDetails)

Undocumented member.

getOrganizationsAccessReportResponse_httpStatus :: Lens' GetOrganizationsAccessReportResponse Int

The response's http status code.

getOrganizationsAccessReportResponse_jobStatus :: Lens' GetOrganizationsAccessReportResponse JobStatusType

The status of the job.

getOrganizationsAccessReportResponse_jobCreationDate :: Lens' GetOrganizationsAccessReportResponse UTCTime

The date and time, in ISO 8601 date-time format, when the report job was created.