Network.AWS.IAM.GetServerCertificate

Description

Retrieves information about the specified server certificate stored in IAM.

For more information about working with server certificates, see Working with server certificates in the IAM User Guide. This topic includes a list of Amazon Web Services services that can use the server certificates that you manage with IAM.

Creating a Request

data GetServerCertificate

See: newGetServerCertificate smart constructor.

Constructors

GetServerCertificate'

Fields

serverCertificateName :: Text

The name of the server certificate you want to retrieve information about.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

newGetServerCertificate

Arguments

:: Text
$sel:serverCertificateName:GetServerCertificate'

-> GetServerCertificate

Create a value of GetServerCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serverCertificateName:GetServerCertificate', getServerCertificate_serverCertificateName - The name of the server certificate you want to retrieve information about.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Request Lenses

getServerCertificate_serverCertificateName :: Lens' GetServerCertificate Text

The name of the server certificate you want to retrieve information about.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Destructuring the Response

data GetServerCertificateResponse

Contains the response to a successful GetServerCertificate request.

See: newGetServerCertificateResponse smart constructor.

Constructors

GetServerCertificateResponse'

Fields

httpStatus :: Int

The response's http status code.

serverCertificate :: ServerCertificate

A structure containing details about the server certificate.

newGetServerCertificateResponse

Arguments

:: Int
$sel:httpStatus:GetServerCertificateResponse'

-> ServerCertificate
$sel:serverCertificate:GetServerCertificateResponse'

-> GetServerCertificateResponse

Create a value of GetServerCertificateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetServerCertificateResponse', getServerCertificateResponse_httpStatus - The response's http status code.

$sel:serverCertificate:GetServerCertificateResponse', getServerCertificateResponse_serverCertificate - A structure containing details about the server certificate.

Response Lenses

getServerCertificateResponse_httpStatus :: Lens' GetServerCertificateResponse Int

The response's http status code.

getServerCertificateResponse_serverCertificate :: Lens' GetServerCertificateResponse ServerCertificate

A structure containing details about the server certificate.