Network.AWS.IAM.GetServiceLastAccessedDetailsWithEntities

Description

After you generate a group or policy report using the GenerateServiceLastAccessedDetails operation, you can use the JobId parameter in GetServiceLastAccessedDetailsWithEntities. This operation retrieves the status of your report job and a list of entities that could have used group or policy permissions to access the specified service.

Group – For a group report, this operation returns a list of users in the group that could have used the group’s policies in an attempt to access the service.

Policy – For a policy report, this operation returns a list of entities (users or roles) that could have used the policy in an attempt to access the service.

You can also use this operation for user or role reports to retrieve details about those entities.

If the operation fails, the GetServiceLastAccessedDetailsWithEntities operation returns the reason that it failed.

By default, the list of associated entities is sorted by date, with the most recent access listed first.

Creating a Request

data GetServiceLastAccessedDetailsWithEntities

See: newGetServiceLastAccessedDetailsWithEntities smart constructor.

Constructors

GetServiceLastAccessedDetailsWithEntities'

Fields

maxItems :: Maybe Natural

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

marker :: Maybe Text

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

jobId :: Text

The ID of the request generated by the GenerateServiceLastAccessedDetails operation.

serviceNamespace :: Text

The service namespace for an Amazon Web Services service. Provide the service namespace to learn when the IAM entity last attempted to access the specified service.

To learn the service namespace for a service, see Actions, resources, and condition keys for Amazon Web Services services in the IAM User Guide. Choose the name of the service to view details for that service. In the first paragraph, find the service prefix. For example, (service prefix: a4b). For more information about service namespaces, see Amazon Web Services service namespaces in the Amazon Web Services General Reference.

newGetServiceLastAccessedDetailsWithEntities

Arguments

:: Text
$sel:jobId:GetServiceLastAccessedDetailsWithEntities'

-> Text
$sel:serviceNamespace:GetServiceLastAccessedDetailsWithEntities'

-> GetServiceLastAccessedDetailsWithEntities

Create a value of GetServiceLastAccessedDetailsWithEntities with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxItems:GetServiceLastAccessedDetailsWithEntities', getServiceLastAccessedDetailsWithEntities_maxItems - Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

$sel:marker:GetServiceLastAccessedDetailsWithEntities', getServiceLastAccessedDetailsWithEntities_marker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

$sel:jobId:GetServiceLastAccessedDetailsWithEntities', getServiceLastAccessedDetailsWithEntities_jobId - The ID of the request generated by the GenerateServiceLastAccessedDetails operation.

$sel:serviceNamespace:GetServiceLastAccessedDetailsWithEntities', getServiceLastAccessedDetailsWithEntities_serviceNamespace - The service namespace for an Amazon Web Services service. Provide the service namespace to learn when the IAM entity last attempted to access the specified service.

To learn the service namespace for a service, see Actions, resources, and condition keys for Amazon Web Services services in the IAM User Guide. Choose the name of the service to view details for that service. In the first paragraph, find the service prefix. For example, (service prefix: a4b). For more information about service namespaces, see Amazon Web Services service namespaces in the Amazon Web Services General Reference.

Request Lenses

getServiceLastAccessedDetailsWithEntities_maxItems :: Lens' GetServiceLastAccessedDetailsWithEntities (Maybe Natural)

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

getServiceLastAccessedDetailsWithEntities_marker :: Lens' GetServiceLastAccessedDetailsWithEntities (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

getServiceLastAccessedDetailsWithEntities_jobId :: Lens' GetServiceLastAccessedDetailsWithEntities Text

The ID of the request generated by the GenerateServiceLastAccessedDetails operation.

getServiceLastAccessedDetailsWithEntities_serviceNamespace :: Lens' GetServiceLastAccessedDetailsWithEntities Text

The service namespace for an Amazon Web Services service. Provide the service namespace to learn when the IAM entity last attempted to access the specified service.

To learn the service namespace for a service, see Actions, resources, and condition keys for Amazon Web Services services in the IAM User Guide. Choose the name of the service to view details for that service. In the first paragraph, find the service prefix. For example, (service prefix: a4b). For more information about service namespaces, see Amazon Web Services service namespaces in the Amazon Web Services General Reference.

Destructuring the Response

data GetServiceLastAccessedDetailsWithEntitiesResponse

See: newGetServiceLastAccessedDetailsWithEntitiesResponse smart constructor.

Constructors

GetServiceLastAccessedDetailsWithEntitiesResponse'

Fields

isTruncated :: Maybe Bool

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

error :: Maybe ErrorDetails

An object that contains details about the reason the operation failed.

marker :: Maybe Text

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

httpStatus :: Int

The response's http status code.

jobStatus :: JobStatusType

The status of the job.

jobCreationDate :: ISO8601

The date and time, in ISO 8601 date-time format, when the report job was created.

jobCompletionDate :: ISO8601

The date and time, in ISO 8601 date-time format, when the generated report job was completed or failed.

This field is null if the job is still in progress, as indicated by a job status value of IN_PROGRESS.

entityDetailsList :: [EntityDetails]

An EntityDetailsList object that contains details about when an IAM entity (user or role) used group or policy permissions in an attempt to access the specified Amazon Web Services service.

newGetServiceLastAccessedDetailsWithEntitiesResponse

Arguments

:: Int
$sel:httpStatus:GetServiceLastAccessedDetailsWithEntitiesResponse'

-> JobStatusType
$sel:jobStatus:GetServiceLastAccessedDetailsWithEntitiesResponse'

-> UTCTime
$sel:jobCreationDate:GetServiceLastAccessedDetailsWithEntitiesResponse'

-> UTCTime
$sel:jobCompletionDate:GetServiceLastAccessedDetailsWithEntitiesResponse'

-> GetServiceLastAccessedDetailsWithEntitiesResponse

Create a value of GetServiceLastAccessedDetailsWithEntitiesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isTruncated:GetServiceLastAccessedDetailsWithEntitiesResponse', getServiceLastAccessedDetailsWithEntitiesResponse_isTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

$sel:error:GetServiceLastAccessedDetailsWithEntitiesResponse', getServiceLastAccessedDetailsWithEntitiesResponse_error - An object that contains details about the reason the operation failed.

$sel:marker:GetServiceLastAccessedDetailsWithEntities', getServiceLastAccessedDetailsWithEntitiesResponse_marker - When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

$sel:httpStatus:GetServiceLastAccessedDetailsWithEntitiesResponse', getServiceLastAccessedDetailsWithEntitiesResponse_httpStatus - The response's http status code.

$sel:jobStatus:GetServiceLastAccessedDetailsWithEntitiesResponse', getServiceLastAccessedDetailsWithEntitiesResponse_jobStatus - The status of the job.

$sel:jobCreationDate:GetServiceLastAccessedDetailsWithEntitiesResponse', getServiceLastAccessedDetailsWithEntitiesResponse_jobCreationDate - The date and time, in ISO 8601 date-time format, when the report job was created.

$sel:jobCompletionDate:GetServiceLastAccessedDetailsWithEntitiesResponse', getServiceLastAccessedDetailsWithEntitiesResponse_jobCompletionDate - The date and time, in ISO 8601 date-time format, when the generated report job was completed or failed.

This field is null if the job is still in progress, as indicated by a job status value of IN_PROGRESS.

$sel:entityDetailsList:GetServiceLastAccessedDetailsWithEntitiesResponse', getServiceLastAccessedDetailsWithEntitiesResponse_entityDetailsList - An EntityDetailsList object that contains details about when an IAM entity (user or role) used group or policy permissions in an attempt to access the specified Amazon Web Services service.

Response Lenses

getServiceLastAccessedDetailsWithEntitiesResponse_isTruncated :: Lens' GetServiceLastAccessedDetailsWithEntitiesResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

getServiceLastAccessedDetailsWithEntitiesResponse_error :: Lens' GetServiceLastAccessedDetailsWithEntitiesResponse (Maybe ErrorDetails)

An object that contains details about the reason the operation failed.

getServiceLastAccessedDetailsWithEntitiesResponse_marker :: Lens' GetServiceLastAccessedDetailsWithEntitiesResponse (Maybe Text)

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

getServiceLastAccessedDetailsWithEntitiesResponse_httpStatus :: Lens' GetServiceLastAccessedDetailsWithEntitiesResponse Int

The response's http status code.

getServiceLastAccessedDetailsWithEntitiesResponse_jobStatus :: Lens' GetServiceLastAccessedDetailsWithEntitiesResponse JobStatusType

The status of the job.

getServiceLastAccessedDetailsWithEntitiesResponse_jobCreationDate :: Lens' GetServiceLastAccessedDetailsWithEntitiesResponse UTCTime

The date and time, in ISO 8601 date-time format, when the report job was created.

getServiceLastAccessedDetailsWithEntitiesResponse_jobCompletionDate :: Lens' GetServiceLastAccessedDetailsWithEntitiesResponse UTCTime

The date and time, in ISO 8601 date-time format, when the generated report job was completed or failed.

This field is null if the job is still in progress, as indicated by a job status value of IN_PROGRESS.

getServiceLastAccessedDetailsWithEntitiesResponse_entityDetailsList :: Lens' GetServiceLastAccessedDetailsWithEntitiesResponse [EntityDetails]

An EntityDetailsList object that contains details about when an IAM entity (user or role) used group or policy permissions in an attempt to access the specified Amazon Web Services service.