Network.AWS.IAM.ListAttachedGroupPolicies

Description

Lists all managed policies that are attached to the specified IAM group.

An IAM group can also have inline policies embedded with it. To list the inline policies for a group, use ListGroupPolicies. For information about policies, see Managed policies and inline policies in the IAM User Guide.

You can paginate the results using the MaxItems and Marker parameters. You can use the PathPrefix parameter to limit the list of policies to only those matching the specified path prefix. If there are no policies attached to the specified group (or none that match the specified path prefix), the operation returns an empty list.

This operation returns paginated results.

Creating a Request

data ListAttachedGroupPolicies

See: newListAttachedGroupPolicies smart constructor.

Constructors

ListAttachedGroupPolicies'

Fields

pathPrefix :: Maybe Text

The path prefix for filtering the results. This parameter is optional. If it is not included, it defaults to a slash (/), listing all policies.

This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

marker :: Maybe Text

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

maxItems :: Maybe Natural

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

groupName :: Text

The name (friendly name, not ARN) of the group to list attached policies for.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

newListAttachedGroupPolicies

Arguments

:: Text
$sel:groupName:ListAttachedGroupPolicies'

-> ListAttachedGroupPolicies

Create a value of ListAttachedGroupPolicies with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pathPrefix:ListAttachedGroupPolicies', listAttachedGroupPolicies_pathPrefix - The path prefix for filtering the results. This parameter is optional. If it is not included, it defaults to a slash (/), listing all policies.

This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

$sel:marker:ListAttachedGroupPolicies', listAttachedGroupPolicies_marker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

$sel:maxItems:ListAttachedGroupPolicies', listAttachedGroupPolicies_maxItems - Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

$sel:groupName:ListAttachedGroupPolicies', listAttachedGroupPolicies_groupName - The name (friendly name, not ARN) of the group to list attached policies for.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Request Lenses

listAttachedGroupPolicies_pathPrefix :: Lens' ListAttachedGroupPolicies (Maybe Text)

The path prefix for filtering the results. This parameter is optional. If it is not included, it defaults to a slash (/), listing all policies.

This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

listAttachedGroupPolicies_marker :: Lens' ListAttachedGroupPolicies (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

listAttachedGroupPolicies_maxItems :: Lens' ListAttachedGroupPolicies (Maybe Natural)

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

listAttachedGroupPolicies_groupName :: Lens' ListAttachedGroupPolicies Text

The name (friendly name, not ARN) of the group to list attached policies for.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Destructuring the Response

data ListAttachedGroupPoliciesResponse

Contains the response to a successful ListAttachedGroupPolicies request.

See: newListAttachedGroupPoliciesResponse smart constructor.

Constructors

ListAttachedGroupPoliciesResponse'

Fields

attachedPolicies :: Maybe [AttachedPolicy]

A list of the attached policies.

marker :: Maybe Text

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

isTruncated :: Maybe Bool

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

httpStatus :: Int

The response's http status code.

newListAttachedGroupPoliciesResponse

Arguments

:: Int
$sel:httpStatus:ListAttachedGroupPoliciesResponse'

-> ListAttachedGroupPoliciesResponse

Create a value of ListAttachedGroupPoliciesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attachedPolicies:ListAttachedGroupPoliciesResponse', listAttachedGroupPoliciesResponse_attachedPolicies - A list of the attached policies.

$sel:marker:ListAttachedGroupPolicies', listAttachedGroupPoliciesResponse_marker - When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

$sel:isTruncated:ListAttachedGroupPoliciesResponse', listAttachedGroupPoliciesResponse_isTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

$sel:httpStatus:ListAttachedGroupPoliciesResponse', listAttachedGroupPoliciesResponse_httpStatus - The response's http status code.

Response Lenses

listAttachedGroupPoliciesResponse_attachedPolicies :: Lens' ListAttachedGroupPoliciesResponse (Maybe [AttachedPolicy])

A list of the attached policies.

listAttachedGroupPoliciesResponse_marker :: Lens' ListAttachedGroupPoliciesResponse (Maybe Text)

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

listAttachedGroupPoliciesResponse_isTruncated :: Lens' ListAttachedGroupPoliciesResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

listAttachedGroupPoliciesResponse_httpStatus :: Lens' ListAttachedGroupPoliciesResponse Int

The response's http status code.