Network.AWS.IAM.ListAttachedUserPolicies

Description

Lists all managed policies that are attached to the specified IAM user.

An IAM user can also have inline policies embedded with it. To list the inline policies for a user, use ListUserPolicies. For information about policies, see Managed policies and inline policies in the IAM User Guide.

You can paginate the results using the MaxItems and Marker parameters. You can use the PathPrefix parameter to limit the list of policies to only those matching the specified path prefix. If there are no policies attached to the specified group (or none that match the specified path prefix), the operation returns an empty list.

This operation returns paginated results.

Creating a Request

data ListAttachedUserPolicies

See: newListAttachedUserPolicies smart constructor.

Constructors

ListAttachedUserPolicies'

Fields

pathPrefix :: Maybe Text

The path prefix for filtering the results. This parameter is optional. If it is not included, it defaults to a slash (/), listing all policies.

This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

marker :: Maybe Text

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

maxItems :: Maybe Natural

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

userName :: Text

The name (friendly name, not ARN) of the user to list attached policies for.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

newListAttachedUserPolicies

Arguments

:: Text
$sel:userName:ListAttachedUserPolicies'

-> ListAttachedUserPolicies

Create a value of ListAttachedUserPolicies with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pathPrefix:ListAttachedUserPolicies', listAttachedUserPolicies_pathPrefix - The path prefix for filtering the results. This parameter is optional. If it is not included, it defaults to a slash (/), listing all policies.

This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

$sel:marker:ListAttachedUserPolicies', listAttachedUserPolicies_marker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

$sel:maxItems:ListAttachedUserPolicies', listAttachedUserPolicies_maxItems - Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

$sel:userName:ListAttachedUserPolicies', listAttachedUserPolicies_userName - The name (friendly name, not ARN) of the user to list attached policies for.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Request Lenses

listAttachedUserPolicies_pathPrefix :: Lens' ListAttachedUserPolicies (Maybe Text)

The path prefix for filtering the results. This parameter is optional. If it is not included, it defaults to a slash (/), listing all policies.

This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

listAttachedUserPolicies_marker :: Lens' ListAttachedUserPolicies (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

listAttachedUserPolicies_maxItems :: Lens' ListAttachedUserPolicies (Maybe Natural)

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

listAttachedUserPolicies_userName :: Lens' ListAttachedUserPolicies Text

The name (friendly name, not ARN) of the user to list attached policies for.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Destructuring the Response

data ListAttachedUserPoliciesResponse

Contains the response to a successful ListAttachedUserPolicies request.

See: newListAttachedUserPoliciesResponse smart constructor.

Constructors

ListAttachedUserPoliciesResponse'

Fields

attachedPolicies :: Maybe [AttachedPolicy]

A list of the attached policies.

marker :: Maybe Text

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

isTruncated :: Maybe Bool

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

httpStatus :: Int

The response's http status code.

newListAttachedUserPoliciesResponse

Arguments

:: Int
$sel:httpStatus:ListAttachedUserPoliciesResponse'

-> ListAttachedUserPoliciesResponse

Create a value of ListAttachedUserPoliciesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attachedPolicies:ListAttachedUserPoliciesResponse', listAttachedUserPoliciesResponse_attachedPolicies - A list of the attached policies.

$sel:marker:ListAttachedUserPolicies', listAttachedUserPoliciesResponse_marker - When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

$sel:isTruncated:ListAttachedUserPoliciesResponse', listAttachedUserPoliciesResponse_isTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

$sel:httpStatus:ListAttachedUserPoliciesResponse', listAttachedUserPoliciesResponse_httpStatus - The response's http status code.

Response Lenses

listAttachedUserPoliciesResponse_attachedPolicies :: Lens' ListAttachedUserPoliciesResponse (Maybe [AttachedPolicy])

A list of the attached policies.

listAttachedUserPoliciesResponse_marker :: Lens' ListAttachedUserPoliciesResponse (Maybe Text)

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

listAttachedUserPoliciesResponse_isTruncated :: Lens' ListAttachedUserPoliciesResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

listAttachedUserPoliciesResponse_httpStatus :: Lens' ListAttachedUserPoliciesResponse Int

The response's http status code.