Network.AWS.IAM.ListEntitiesForPolicy

Description

Lists all IAM users, groups, and roles that the specified managed policy is attached to.

You can use the optional EntityFilter parameter to limit the results to a particular type of entity (users, groups, or roles). For example, to list only the roles that are attached to the specified policy, set EntityFilter to Role.

You can paginate the results using the MaxItems and Marker parameters.

This operation returns paginated results.

Creating a Request

data ListEntitiesForPolicy

See: newListEntitiesForPolicy smart constructor.

Constructors

ListEntitiesForPolicy'

Fields

pathPrefix :: Maybe Text

The path prefix for filtering the results. This parameter is optional. If it is not included, it defaults to a slash (/), listing all entities.

This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

entityFilter :: Maybe EntityType

The entity type to use for filtering the results.

For example, when EntityFilter is Role, only the roles that are attached to the specified policy are returned. This parameter is optional. If it is not included, all attached entities (users, groups, and roles) are returned. The argument for this parameter must be one of the valid values listed below.

marker :: Maybe Text

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

maxItems :: Maybe Natural

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

policyUsageFilter :: Maybe PolicyUsageType

The policy usage method to use for filtering the results.

To list only permissions policies, set PolicyUsageFilter to PermissionsPolicy. To list only the policies used to set permissions boundaries, set the value to PermissionsBoundary.

This parameter is optional. If it is not included, all policies are returned.

policyArn :: Text

The Amazon Resource Name (ARN) of the IAM policy for which you want the versions.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

newListEntitiesForPolicy

Arguments

:: Text
$sel:policyArn:ListEntitiesForPolicy'

-> ListEntitiesForPolicy

Create a value of ListEntitiesForPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pathPrefix:ListEntitiesForPolicy', listEntitiesForPolicy_pathPrefix - The path prefix for filtering the results. This parameter is optional. If it is not included, it defaults to a slash (/), listing all entities.

This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

$sel:entityFilter:ListEntitiesForPolicy', listEntitiesForPolicy_entityFilter - The entity type to use for filtering the results.

For example, when EntityFilter is Role, only the roles that are attached to the specified policy are returned. This parameter is optional. If it is not included, all attached entities (users, groups, and roles) are returned. The argument for this parameter must be one of the valid values listed below.

$sel:marker:ListEntitiesForPolicy', listEntitiesForPolicy_marker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

$sel:maxItems:ListEntitiesForPolicy', listEntitiesForPolicy_maxItems - Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

$sel:policyUsageFilter:ListEntitiesForPolicy', listEntitiesForPolicy_policyUsageFilter - The policy usage method to use for filtering the results.

To list only permissions policies, set PolicyUsageFilter to PermissionsPolicy. To list only the policies used to set permissions boundaries, set the value to PermissionsBoundary.

This parameter is optional. If it is not included, all policies are returned.

$sel:policyArn:ListEntitiesForPolicy', listEntitiesForPolicy_policyArn - The Amazon Resource Name (ARN) of the IAM policy for which you want the versions.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

Request Lenses

listEntitiesForPolicy_pathPrefix :: Lens' ListEntitiesForPolicy (Maybe Text)

The path prefix for filtering the results. This parameter is optional. If it is not included, it defaults to a slash (/), listing all entities.

This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

listEntitiesForPolicy_entityFilter :: Lens' ListEntitiesForPolicy (Maybe EntityType)

The entity type to use for filtering the results.

For example, when EntityFilter is Role, only the roles that are attached to the specified policy are returned. This parameter is optional. If it is not included, all attached entities (users, groups, and roles) are returned. The argument for this parameter must be one of the valid values listed below.

listEntitiesForPolicy_marker :: Lens' ListEntitiesForPolicy (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

listEntitiesForPolicy_maxItems :: Lens' ListEntitiesForPolicy (Maybe Natural)

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

listEntitiesForPolicy_policyUsageFilter :: Lens' ListEntitiesForPolicy (Maybe PolicyUsageType)

The policy usage method to use for filtering the results.

To list only permissions policies, set PolicyUsageFilter to PermissionsPolicy. To list only the policies used to set permissions boundaries, set the value to PermissionsBoundary.

This parameter is optional. If it is not included, all policies are returned.

listEntitiesForPolicy_policyArn :: Lens' ListEntitiesForPolicy Text

The Amazon Resource Name (ARN) of the IAM policy for which you want the versions.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

Destructuring the Response

data ListEntitiesForPolicyResponse

Contains the response to a successful ListEntitiesForPolicy request.

See: newListEntitiesForPolicyResponse smart constructor.

Constructors

ListEntitiesForPolicyResponse'

Fields

policyGroups :: Maybe [PolicyGroup]

A list of IAM groups that the policy is attached to.

policyRoles :: Maybe [PolicyRole]

A list of IAM roles that the policy is attached to.

marker :: Maybe Text

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

policyUsers :: Maybe [PolicyUser]

A list of IAM users that the policy is attached to.

isTruncated :: Maybe Bool

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

httpStatus :: Int

The response's http status code.

newListEntitiesForPolicyResponse

Arguments

:: Int
$sel:httpStatus:ListEntitiesForPolicyResponse'

-> ListEntitiesForPolicyResponse

Create a value of ListEntitiesForPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyGroups:ListEntitiesForPolicyResponse', listEntitiesForPolicyResponse_policyGroups - A list of IAM groups that the policy is attached to.

$sel:policyRoles:ListEntitiesForPolicyResponse', listEntitiesForPolicyResponse_policyRoles - A list of IAM roles that the policy is attached to.

$sel:marker:ListEntitiesForPolicy', listEntitiesForPolicyResponse_marker - When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

$sel:policyUsers:ListEntitiesForPolicyResponse', listEntitiesForPolicyResponse_policyUsers - A list of IAM users that the policy is attached to.

$sel:isTruncated:ListEntitiesForPolicyResponse', listEntitiesForPolicyResponse_isTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

$sel:httpStatus:ListEntitiesForPolicyResponse', listEntitiesForPolicyResponse_httpStatus - The response's http status code.

Response Lenses

listEntitiesForPolicyResponse_policyGroups :: Lens' ListEntitiesForPolicyResponse (Maybe [PolicyGroup])

A list of IAM groups that the policy is attached to.

listEntitiesForPolicyResponse_policyRoles :: Lens' ListEntitiesForPolicyResponse (Maybe [PolicyRole])

A list of IAM roles that the policy is attached to.

listEntitiesForPolicyResponse_marker :: Lens' ListEntitiesForPolicyResponse (Maybe Text)

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

listEntitiesForPolicyResponse_policyUsers :: Lens' ListEntitiesForPolicyResponse (Maybe [PolicyUser])

A list of IAM users that the policy is attached to.

listEntitiesForPolicyResponse_isTruncated :: Lens' ListEntitiesForPolicyResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

listEntitiesForPolicyResponse_httpStatus :: Lens' ListEntitiesForPolicyResponse Int

The response's http status code.