Network.AWS.IAM.ListInstanceProfiles

Description

Lists the instance profiles that have the specified path prefix. If there are none, the operation returns an empty list. For more information about instance profiles, see About instance profiles.

IAM resource-listing operations return a subset of the available attributes for the resource. For example, this operation does not return tags, even though they are an attribute of the returned object. To view all of the information for an instance profile, see GetInstanceProfile.

You can paginate the results using the MaxItems and Marker parameters.

This operation returns paginated results.

Creating a Request

data ListInstanceProfiles

See: newListInstanceProfiles smart constructor.

Constructors

ListInstanceProfiles'

Fields

pathPrefix :: Maybe Text

The path prefix for filtering the results. For example, the prefix /application_abc/component_xyz/ gets all instance profiles whose path starts with /application_abc/component_xyz/.

This parameter is optional. If it is not included, it defaults to a slash (/), listing all instance profiles. This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

maxItems :: Maybe Natural

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

marker :: Maybe Text

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

newListInstanceProfiles :: ListInstanceProfiles

Create a value of ListInstanceProfiles with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pathPrefix:ListInstanceProfiles', listInstanceProfiles_pathPrefix - The path prefix for filtering the results. For example, the prefix /application_abc/component_xyz/ gets all instance profiles whose path starts with /application_abc/component_xyz/.

This parameter is optional. If it is not included, it defaults to a slash (/), listing all instance profiles. This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

$sel:maxItems:ListInstanceProfiles', listInstanceProfiles_maxItems - Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

$sel:marker:ListInstanceProfiles', listInstanceProfiles_marker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

Request Lenses

listInstanceProfiles_pathPrefix :: Lens' ListInstanceProfiles (Maybe Text)

The path prefix for filtering the results. For example, the prefix /application_abc/component_xyz/ gets all instance profiles whose path starts with /application_abc/component_xyz/.

This parameter is optional. If it is not included, it defaults to a slash (/), listing all instance profiles. This parameter allows (through its regex pattern) a string of characters consisting of either a forward slash (/) by itself or a string that must begin and end with forward slashes. In addition, it can contain any ASCII character from the ! (\u0021) through the DEL character (\u007F), including most punctuation characters, digits, and upper and lowercased letters.

listInstanceProfiles_maxItems :: Lens' ListInstanceProfiles (Maybe Natural)

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

listInstanceProfiles_marker :: Lens' ListInstanceProfiles (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

Destructuring the Response

data ListInstanceProfilesResponse

Contains the response to a successful ListInstanceProfiles request.

See: newListInstanceProfilesResponse smart constructor.

Constructors

ListInstanceProfilesResponse'

Fields

isTruncated :: Maybe Bool

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

marker :: Maybe Text

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

httpStatus :: Int

The response's http status code.

instanceProfiles :: [InstanceProfile]

A list of instance profiles.

newListInstanceProfilesResponse

Arguments

:: Int
$sel:httpStatus:ListInstanceProfilesResponse'

-> ListInstanceProfilesResponse

Create a value of ListInstanceProfilesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isTruncated:ListInstanceProfilesResponse', listInstanceProfilesResponse_isTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

$sel:marker:ListInstanceProfiles', listInstanceProfilesResponse_marker - When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

$sel:httpStatus:ListInstanceProfilesResponse', listInstanceProfilesResponse_httpStatus - The response's http status code.

$sel:instanceProfiles:ListInstanceProfilesResponse', listInstanceProfilesResponse_instanceProfiles - A list of instance profiles.

Response Lenses

listInstanceProfilesResponse_isTruncated :: Lens' ListInstanceProfilesResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

listInstanceProfilesResponse_marker :: Lens' ListInstanceProfilesResponse (Maybe Text)

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

listInstanceProfilesResponse_httpStatus :: Lens' ListInstanceProfilesResponse Int

The response's http status code.

listInstanceProfilesResponse_instanceProfiles :: Lens' ListInstanceProfilesResponse [InstanceProfile]

A list of instance profiles.