Network.AWS.IAM.ListOpenIDConnectProviders

Description

Lists information about the IAM OpenID Connect (OIDC) provider resource objects defined in the Amazon Web Services account.

IAM resource-listing operations return a subset of the available attributes for the resource. For example, this operation does not return tags, even though they are an attribute of the returned object. To view all of the information for an OIDC provider, see GetOpenIDConnectProvider.

Creating a Request

data ListOpenIDConnectProviders

See: newListOpenIDConnectProviders smart constructor.

Constructors

ListOpenIDConnectProviders'

newListOpenIDConnectProviders :: ListOpenIDConnectProviders

Create a value of ListOpenIDConnectProviders with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data ListOpenIDConnectProvidersResponse

Contains the response to a successful ListOpenIDConnectProviders request.

See: newListOpenIDConnectProvidersResponse smart constructor.

Constructors

ListOpenIDConnectProvidersResponse'

Fields

openIDConnectProviderList :: Maybe [OpenIDConnectProviderListEntry]

The list of IAM OIDC provider resource objects defined in the Amazon Web Services account.

httpStatus :: Int

The response's http status code.

newListOpenIDConnectProvidersResponse

Arguments

:: Int
$sel:httpStatus:ListOpenIDConnectProvidersResponse'

-> ListOpenIDConnectProvidersResponse

Create a value of ListOpenIDConnectProvidersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:openIDConnectProviderList:ListOpenIDConnectProvidersResponse', listOpenIDConnectProvidersResponse_openIDConnectProviderList - The list of IAM OIDC provider resource objects defined in the Amazon Web Services account.

$sel:httpStatus:ListOpenIDConnectProvidersResponse', listOpenIDConnectProvidersResponse_httpStatus - The response's http status code.

Response Lenses

listOpenIDConnectProvidersResponse_openIDConnectProviderList :: Lens' ListOpenIDConnectProvidersResponse (Maybe [OpenIDConnectProviderListEntry])

The list of IAM OIDC provider resource objects defined in the Amazon Web Services account.

listOpenIDConnectProvidersResponse_httpStatus :: Lens' ListOpenIDConnectProvidersResponse Int

The response's http status code.