Network.AWS.IAM.ListPoliciesGrantingServiceAccess

Description

Retrieves a list of policies that the IAM identity (user, group, or role) can use to access each specified service.

This operation does not use other policy types when determining whether a resource could access a service. These other policy types include resource-based policies, access control lists, Organizations policies, IAM permissions boundaries, and STS assume role policies. It only applies permissions policy logic. For more about the evaluation of policy types, see Evaluating policies in the IAM User Guide.

The list of policies returned by the operation depends on the ARN of the identity that you provide.

User – The list of policies includes the managed and inline policies that are attached to the user directly. The list also includes any additional managed and inline policies that are attached to the group to which the user belongs.

Group – The list of policies includes only the managed and inline policies that are attached to the group directly. Policies that are attached to the group’s user are not included.

Role – The list of policies includes only the managed and inline policies that are attached to the role.

For each managed policy, this operation returns the ARN and policy name. For each inline policy, it returns the policy name and the entity to which it is attached. Inline policies do not have an ARN. For more information about these policy types, see Managed policies and inline policies in the IAM User Guide.

Policies that are attached to users and roles as permissions boundaries are not returned. To view which managed policy is currently used to set the permissions boundary for a user or role, use the GetUser or GetRole operations.

Creating a Request

data ListPoliciesGrantingServiceAccess

See: newListPoliciesGrantingServiceAccess smart constructor.

Constructors

ListPoliciesGrantingServiceAccess'

Fields

marker :: Maybe Text

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

arn :: Text

The ARN of the IAM identity (user, group, or role) whose policies you want to list.

serviceNamespaces :: NonEmpty Text

The service namespace for the Amazon Web Services services whose policies you want to list.

To learn the service namespace for a service, see Actions, resources, and condition keys for Amazon Web Services services in the IAM User Guide. Choose the name of the service to view details for that service. In the first paragraph, find the service prefix. For example, (service prefix: a4b). For more information about service namespaces, see Amazon Web Services service namespaces in the Amazon Web Services General Reference.

newListPoliciesGrantingServiceAccess

Arguments

:: Text
$sel:arn:ListPoliciesGrantingServiceAccess'

-> NonEmpty Text
$sel:serviceNamespaces:ListPoliciesGrantingServiceAccess'

-> ListPoliciesGrantingServiceAccess

Create a value of ListPoliciesGrantingServiceAccess with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:ListPoliciesGrantingServiceAccess', listPoliciesGrantingServiceAccess_marker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

$sel:arn:ListPoliciesGrantingServiceAccess', listPoliciesGrantingServiceAccess_arn - The ARN of the IAM identity (user, group, or role) whose policies you want to list.

$sel:serviceNamespaces:ListPoliciesGrantingServiceAccess', listPoliciesGrantingServiceAccess_serviceNamespaces - The service namespace for the Amazon Web Services services whose policies you want to list.

To learn the service namespace for a service, see Actions, resources, and condition keys for Amazon Web Services services in the IAM User Guide. Choose the name of the service to view details for that service. In the first paragraph, find the service prefix. For example, (service prefix: a4b). For more information about service namespaces, see Amazon Web Services service namespaces in the Amazon Web Services General Reference.

Request Lenses

listPoliciesGrantingServiceAccess_marker :: Lens' ListPoliciesGrantingServiceAccess (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

listPoliciesGrantingServiceAccess_arn :: Lens' ListPoliciesGrantingServiceAccess Text

The ARN of the IAM identity (user, group, or role) whose policies you want to list.

listPoliciesGrantingServiceAccess_serviceNamespaces :: Lens' ListPoliciesGrantingServiceAccess (NonEmpty Text)

The service namespace for the Amazon Web Services services whose policies you want to list.

To learn the service namespace for a service, see Actions, resources, and condition keys for Amazon Web Services services in the IAM User Guide. Choose the name of the service to view details for that service. In the first paragraph, find the service prefix. For example, (service prefix: a4b). For more information about service namespaces, see Amazon Web Services service namespaces in the Amazon Web Services General Reference.

Destructuring the Response

data ListPoliciesGrantingServiceAccessResponse

See: newListPoliciesGrantingServiceAccessResponse smart constructor.

Constructors

ListPoliciesGrantingServiceAccessResponse'

Fields

isTruncated :: Maybe Bool

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

marker :: Maybe Text

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

httpStatus :: Int

The response's http status code.

policiesGrantingServiceAccess :: [ListPoliciesGrantingServiceAccessEntry]

A ListPoliciesGrantingServiceAccess object that contains details about the permissions policies attached to the specified identity (user, group, or role).

newListPoliciesGrantingServiceAccessResponse

Arguments

:: Int
$sel:httpStatus:ListPoliciesGrantingServiceAccessResponse'

-> ListPoliciesGrantingServiceAccessResponse

Create a value of ListPoliciesGrantingServiceAccessResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isTruncated:ListPoliciesGrantingServiceAccessResponse', listPoliciesGrantingServiceAccessResponse_isTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

$sel:marker:ListPoliciesGrantingServiceAccess', listPoliciesGrantingServiceAccessResponse_marker - When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

$sel:httpStatus:ListPoliciesGrantingServiceAccessResponse', listPoliciesGrantingServiceAccessResponse_httpStatus - The response's http status code.

$sel:policiesGrantingServiceAccess:ListPoliciesGrantingServiceAccessResponse', listPoliciesGrantingServiceAccessResponse_policiesGrantingServiceAccess - A ListPoliciesGrantingServiceAccess object that contains details about the permissions policies attached to the specified identity (user, group, or role).

Response Lenses

listPoliciesGrantingServiceAccessResponse_isTruncated :: Lens' ListPoliciesGrantingServiceAccessResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

listPoliciesGrantingServiceAccessResponse_marker :: Lens' ListPoliciesGrantingServiceAccessResponse (Maybe Text)

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

listPoliciesGrantingServiceAccessResponse_httpStatus :: Lens' ListPoliciesGrantingServiceAccessResponse Int

The response's http status code.

listPoliciesGrantingServiceAccessResponse_policiesGrantingServiceAccess :: Lens' ListPoliciesGrantingServiceAccessResponse [ListPoliciesGrantingServiceAccessEntry]

A ListPoliciesGrantingServiceAccess object that contains details about the permissions policies attached to the specified identity (user, group, or role).