Network.AWS.IAM.ListPolicyVersions

Description

Lists information about the versions of the specified managed policy, including the version that is currently set as the policy's default version.

For more information about managed policies, see Managed policies and inline policies in the IAM User Guide.

This operation returns paginated results.

Creating a Request

data ListPolicyVersions

See: newListPolicyVersions smart constructor.

Constructors

ListPolicyVersions'

Fields

maxItems :: Maybe Natural

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

marker :: Maybe Text

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

policyArn :: Text

The Amazon Resource Name (ARN) of the IAM policy for which you want the versions.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

newListPolicyVersions

Arguments

:: Text
$sel:policyArn:ListPolicyVersions'

-> ListPolicyVersions

Create a value of ListPolicyVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxItems:ListPolicyVersions', listPolicyVersions_maxItems - Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

$sel:marker:ListPolicyVersions', listPolicyVersions_marker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

$sel:policyArn:ListPolicyVersions', listPolicyVersions_policyArn - The Amazon Resource Name (ARN) of the IAM policy for which you want the versions.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

Request Lenses

listPolicyVersions_maxItems :: Lens' ListPolicyVersions (Maybe Natural)

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

listPolicyVersions_marker :: Lens' ListPolicyVersions (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

listPolicyVersions_policyArn :: Lens' ListPolicyVersions Text

The Amazon Resource Name (ARN) of the IAM policy for which you want the versions.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

Destructuring the Response

data ListPolicyVersionsResponse

Contains the response to a successful ListPolicyVersions request.

See: newListPolicyVersionsResponse smart constructor.

Constructors

ListPolicyVersionsResponse'

Fields

versions :: Maybe [PolicyVersion]

A list of policy versions.

For more information about managed policy versions, see Versioning for managed policies in the IAM User Guide.

isTruncated :: Maybe Bool

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

marker :: Maybe Text

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

httpStatus :: Int

The response's http status code.

newListPolicyVersionsResponse

Arguments

:: Int
$sel:httpStatus:ListPolicyVersionsResponse'

-> ListPolicyVersionsResponse

Create a value of ListPolicyVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versions:ListPolicyVersionsResponse', listPolicyVersionsResponse_versions - A list of policy versions.

For more information about managed policy versions, see Versioning for managed policies in the IAM User Guide.

$sel:isTruncated:ListPolicyVersionsResponse', listPolicyVersionsResponse_isTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

$sel:marker:ListPolicyVersions', listPolicyVersionsResponse_marker - When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

$sel:httpStatus:ListPolicyVersionsResponse', listPolicyVersionsResponse_httpStatus - The response's http status code.

Response Lenses

listPolicyVersionsResponse_versions :: Lens' ListPolicyVersionsResponse (Maybe [PolicyVersion])

A list of policy versions.

For more information about managed policy versions, see Versioning for managed policies in the IAM User Guide.

listPolicyVersionsResponse_isTruncated :: Lens' ListPolicyVersionsResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

listPolicyVersionsResponse_marker :: Lens' ListPolicyVersionsResponse (Maybe Text)

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

listPolicyVersionsResponse_httpStatus :: Lens' ListPolicyVersionsResponse Int

The response's http status code.