Network.AWS.IAM.ListUserPolicies

Description

Lists the names of the inline policies embedded in the specified IAM user.

An IAM user can also have managed policies attached to it. To list the managed policies that are attached to a user, use ListAttachedUserPolicies. For more information about policies, see Managed policies and inline policies in the IAM User Guide.

You can paginate the results using the MaxItems and Marker parameters. If there are no inline policies embedded with the specified user, the operation returns an empty list.

This operation returns paginated results.

Creating a Request

data ListUserPolicies

See: newListUserPolicies smart constructor.

Constructors

ListUserPolicies'

Fields

maxItems :: Maybe Natural

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

marker :: Maybe Text

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

userName :: Text

The name of the user to list policies for.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

newListUserPolicies

Arguments

:: Text
$sel:userName:ListUserPolicies'

-> ListUserPolicies

Create a value of ListUserPolicies with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxItems:ListUserPolicies', listUserPolicies_maxItems - Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

$sel:marker:ListUserPolicies', listUserPolicies_marker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

$sel:userName:ListUserPolicies', listUserPolicies_userName - The name of the user to list policies for.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Request Lenses

listUserPolicies_maxItems :: Lens' ListUserPolicies (Maybe Natural)

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

listUserPolicies_marker :: Lens' ListUserPolicies (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

listUserPolicies_userName :: Lens' ListUserPolicies Text

The name of the user to list policies for.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Destructuring the Response

data ListUserPoliciesResponse

Contains the response to a successful ListUserPolicies request.

See: newListUserPoliciesResponse smart constructor.

Constructors

ListUserPoliciesResponse'

Fields

isTruncated :: Maybe Bool

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

marker :: Maybe Text

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

httpStatus :: Int

The response's http status code.

policyNames :: [Text]

A list of policy names.

newListUserPoliciesResponse

Arguments

:: Int
$sel:httpStatus:ListUserPoliciesResponse'

-> ListUserPoliciesResponse

Create a value of ListUserPoliciesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isTruncated:ListUserPoliciesResponse', listUserPoliciesResponse_isTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

$sel:marker:ListUserPolicies', listUserPoliciesResponse_marker - When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

$sel:httpStatus:ListUserPoliciesResponse', listUserPoliciesResponse_httpStatus - The response's http status code.

$sel:policyNames:ListUserPoliciesResponse', listUserPoliciesResponse_policyNames - A list of policy names.

Response Lenses

listUserPoliciesResponse_isTruncated :: Lens' ListUserPoliciesResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

listUserPoliciesResponse_marker :: Lens' ListUserPoliciesResponse (Maybe Text)

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

listUserPoliciesResponse_httpStatus :: Lens' ListUserPoliciesResponse Int

The response's http status code.

listUserPoliciesResponse_policyNames :: Lens' ListUserPoliciesResponse [Text]

A list of policy names.