Network.AWS.IAM.ListVirtualMFADevices

Description

Lists the virtual MFA devices defined in the Amazon Web Services account by assignment status. If you do not specify an assignment status, the operation returns a list of all virtual MFA devices. Assignment status can be Assigned, Unassigned, or Any.

IAM resource-listing operations return a subset of the available attributes for the resource. For example, this operation does not return tags, even though they are an attribute of the returned object. To view all of the information for a virtual MFA device, see ListVirtualMFADevices.

You can paginate the results using the MaxItems and Marker parameters.

This operation returns paginated results.

Creating a Request

data ListVirtualMFADevices

See: newListVirtualMFADevices smart constructor.

Constructors

ListVirtualMFADevices'

Fields

assignmentStatus :: Maybe AssignmentStatusType

The status (Unassigned or Assigned) of the devices to list. If you do not specify an AssignmentStatus, the operation defaults to Any, which lists both assigned and unassigned virtual MFA devices.,

maxItems :: Maybe Natural

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

marker :: Maybe Text

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

newListVirtualMFADevices :: ListVirtualMFADevices

Create a value of ListVirtualMFADevices with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assignmentStatus:ListVirtualMFADevices', listVirtualMFADevices_assignmentStatus - The status (Unassigned or Assigned) of the devices to list. If you do not specify an AssignmentStatus, the operation defaults to Any, which lists both assigned and unassigned virtual MFA devices.,

$sel:maxItems:ListVirtualMFADevices', listVirtualMFADevices_maxItems - Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

$sel:marker:ListVirtualMFADevices', listVirtualMFADevices_marker - Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

Request Lenses

listVirtualMFADevices_assignmentStatus :: Lens' ListVirtualMFADevices (Maybe AssignmentStatusType)

The status (Unassigned or Assigned) of the devices to list. If you do not specify an AssignmentStatus, the operation defaults to Any, which lists both assigned and unassigned virtual MFA devices.,

listVirtualMFADevices_maxItems :: Lens' ListVirtualMFADevices (Maybe Natural)

Use this only when paginating results to indicate the maximum number of items you want in the response. If additional items exist beyond the maximum you specify, the IsTruncated response element is true.

If you do not include this parameter, the number of items defaults to 100. Note that IAM might return fewer results, even when there are more results available. In that case, the IsTruncated response element returns true, and Marker contains a value to include in the subsequent call that tells the service where to continue from.

listVirtualMFADevices_marker :: Lens' ListVirtualMFADevices (Maybe Text)

Use this parameter only when paginating results and only after you receive a response indicating that the results are truncated. Set it to the value of the Marker element in the response that you received to indicate where the next call should start.

Destructuring the Response

data ListVirtualMFADevicesResponse

Contains the response to a successful ListVirtualMFADevices request.

See: newListVirtualMFADevicesResponse smart constructor.

Constructors

ListVirtualMFADevicesResponse'

Fields

isTruncated :: Maybe Bool

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

marker :: Maybe Text

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

httpStatus :: Int

The response's http status code.

virtualMFADevices :: [VirtualMFADevice]

The list of virtual MFA devices in the current account that match the AssignmentStatus value that was passed in the request.

newListVirtualMFADevicesResponse

Arguments

:: Int
$sel:httpStatus:ListVirtualMFADevicesResponse'

-> ListVirtualMFADevicesResponse

Create a value of ListVirtualMFADevicesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isTruncated:ListVirtualMFADevicesResponse', listVirtualMFADevicesResponse_isTruncated - A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

$sel:marker:ListVirtualMFADevices', listVirtualMFADevicesResponse_marker - When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

$sel:httpStatus:ListVirtualMFADevicesResponse', listVirtualMFADevicesResponse_httpStatus - The response's http status code.

$sel:virtualMFADevices:ListVirtualMFADevicesResponse', listVirtualMFADevicesResponse_virtualMFADevices - The list of virtual MFA devices in the current account that match the AssignmentStatus value that was passed in the request.

Response Lenses

listVirtualMFADevicesResponse_isTruncated :: Lens' ListVirtualMFADevicesResponse (Maybe Bool)

A flag that indicates whether there are more items to return. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more items. Note that IAM might return fewer than the MaxItems number of results even when there are more results available. We recommend that you check IsTruncated after every call to ensure that you receive all your results.

listVirtualMFADevicesResponse_marker :: Lens' ListVirtualMFADevicesResponse (Maybe Text)

When IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

listVirtualMFADevicesResponse_httpStatus :: Lens' ListVirtualMFADevicesResponse Int

The response's http status code.

listVirtualMFADevicesResponse_virtualMFADevices :: Lens' ListVirtualMFADevicesResponse [VirtualMFADevice]

The list of virtual MFA devices in the current account that match the AssignmentStatus value that was passed in the request.