Network.AWS.IAM.PutRolePolicy

Description

Adds or updates an inline policy document that is embedded in the specified IAM role.

When you embed an inline policy in a role, the inline policy is used as part of the role's access (permissions) policy. The role's trust policy is created at the same time as the role, using CreateRole. You can update a role's trust policy using UpdateAssumeRolePolicy. For more information about IAM roles, see Using roles to delegate permissions and federate identities.

A role can also have a managed policy attached to it. To attach a managed policy to a role, use AttachRolePolicy. To create a new managed policy, use CreatePolicy. For information about policies, see Managed policies and inline policies in the IAM User Guide.

For information about the maximum number of inline policies that you can embed with a role, see IAM and STS quotas in the IAM User Guide.

Because policy documents can be large, you should use POST rather than GET when calling PutRolePolicy. For general information about using the Query API with IAM, see Making query requests in the IAM User Guide.

Creating a Request

data PutRolePolicy

See: newPutRolePolicy smart constructor.

Constructors

PutRolePolicy'

Fields

roleName :: Text

The name of the role to associate the policy with.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

policyName :: Text

The name of the policy document.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

policyDocument :: Text

The policy document.

You must provide policies in JSON format in IAM. However, for CloudFormation templates formatted in YAML, you can provide the policy in JSON or YAML format. CloudFormation always converts a YAML policy to JSON format before submitting it to IAM.

The regex pattern used to validate this parameter is a string of characters consisting of the following:

Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range

The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)

The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)

newPutRolePolicy

Arguments

:: Text
$sel:roleName:PutRolePolicy'

-> Text
$sel:policyName:PutRolePolicy'

-> Text
$sel:policyDocument:PutRolePolicy'

-> PutRolePolicy

Create a value of PutRolePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleName:PutRolePolicy', putRolePolicy_roleName - The name of the role to associate the policy with.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:policyName:PutRolePolicy', putRolePolicy_policyName - The name of the policy document.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:policyDocument:PutRolePolicy', putRolePolicy_policyDocument - The policy document.

You must provide policies in JSON format in IAM. However, for CloudFormation templates formatted in YAML, you can provide the policy in JSON or YAML format. CloudFormation always converts a YAML policy to JSON format before submitting it to IAM.

The regex pattern used to validate this parameter is a string of characters consisting of the following:

Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range

The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)

The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)

Request Lenses

putRolePolicy_roleName :: Lens' PutRolePolicy Text

The name of the role to associate the policy with.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

putRolePolicy_policyName :: Lens' PutRolePolicy Text

The name of the policy document.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

putRolePolicy_policyDocument :: Lens' PutRolePolicy Text

The policy document.

You must provide policies in JSON format in IAM. However, for CloudFormation templates formatted in YAML, you can provide the policy in JSON or YAML format. CloudFormation always converts a YAML policy to JSON format before submitting it to IAM.

The regex pattern used to validate this parameter is a string of characters consisting of the following:

Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range

The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)

The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)

Destructuring the Response

data PutRolePolicyResponse

See: newPutRolePolicyResponse smart constructor.

Constructors

PutRolePolicyResponse'

newPutRolePolicyResponse :: PutRolePolicyResponse

Create a value of PutRolePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.