Network.AWS.IAM.PutUserPolicy

Description

Adds or updates an inline policy document that is embedded in the specified IAM user.

An IAM user can also have a managed policy attached to it. To attach a managed policy to a user, use AttachUserPolicy. To create a new managed policy, use CreatePolicy. For information about policies, see Managed policies and inline policies in the IAM User Guide.

For information about the maximum number of inline policies that you can embed in a user, see IAM and STS quotas in the IAM User Guide.

Because policy documents can be large, you should use POST rather than GET when calling PutUserPolicy. For general information about using the Query API with IAM, see Making query requests in the IAM User Guide.

Creating a Request

data PutUserPolicy

See: newPutUserPolicy smart constructor.

Constructors

PutUserPolicy'

Fields

userName :: Text

The name of the user to associate the policy with.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

policyName :: Text

The name of the policy document.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

policyDocument :: Text

The policy document.

You must provide policies in JSON format in IAM. However, for CloudFormation templates formatted in YAML, you can provide the policy in JSON or YAML format. CloudFormation always converts a YAML policy to JSON format before submitting it to IAM.

The regex pattern used to validate this parameter is a string of characters consisting of the following:

Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range

The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)

The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)

newPutUserPolicy

Arguments

:: Text
$sel:userName:PutUserPolicy'

-> Text
$sel:policyName:PutUserPolicy'

-> Text
$sel:policyDocument:PutUserPolicy'

-> PutUserPolicy

Create a value of PutUserPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userName:PutUserPolicy', putUserPolicy_userName - The name of the user to associate the policy with.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:policyName:PutUserPolicy', putUserPolicy_policyName - The name of the policy document.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:policyDocument:PutUserPolicy', putUserPolicy_policyDocument - The policy document.

You must provide policies in JSON format in IAM. However, for CloudFormation templates formatted in YAML, you can provide the policy in JSON or YAML format. CloudFormation always converts a YAML policy to JSON format before submitting it to IAM.

The regex pattern used to validate this parameter is a string of characters consisting of the following:

Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range

The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)

The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)

Request Lenses

putUserPolicy_userName :: Lens' PutUserPolicy Text

The name of the user to associate the policy with.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

putUserPolicy_policyName :: Lens' PutUserPolicy Text

The name of the policy document.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

putUserPolicy_policyDocument :: Lens' PutUserPolicy Text

The policy document.

You must provide policies in JSON format in IAM. However, for CloudFormation templates formatted in YAML, you can provide the policy in JSON or YAML format. CloudFormation always converts a YAML policy to JSON format before submitting it to IAM.

The regex pattern used to validate this parameter is a string of characters consisting of the following:

Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range

The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)

The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)

Destructuring the Response

data PutUserPolicyResponse

See: newPutUserPolicyResponse smart constructor.

Constructors

PutUserPolicyResponse'

newPutUserPolicyResponse :: PutUserPolicyResponse

Create a value of PutUserPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.