Network.AWS.IAM.RemoveClientIDFromOpenIDConnectProvider

Description

Removes the specified client ID (also known as audience) from the list of client IDs registered for the specified IAM OpenID Connect (OIDC) provider resource object.

This operation is idempotent; it does not fail or return an error if you try to remove a client ID that does not exist.

Creating a Request

data RemoveClientIDFromOpenIDConnectProvider

See: newRemoveClientIDFromOpenIDConnectProvider smart constructor.

Constructors

RemoveClientIDFromOpenIDConnectProvider'

Fields

openIDConnectProviderArn :: Text

The Amazon Resource Name (ARN) of the IAM OIDC provider resource to remove the client ID from. You can get a list of OIDC provider ARNs by using the ListOpenIDConnectProviders operation.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

clientID :: Text

The client ID (also known as audience) to remove from the IAM OIDC provider resource. For more information about client IDs, see CreateOpenIDConnectProvider.

newRemoveClientIDFromOpenIDConnectProvider

Arguments

:: Text
$sel:openIDConnectProviderArn:RemoveClientIDFromOpenIDConnectProvider'

-> Text
$sel:clientID:RemoveClientIDFromOpenIDConnectProvider'

-> RemoveClientIDFromOpenIDConnectProvider

Create a value of RemoveClientIDFromOpenIDConnectProvider with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:openIDConnectProviderArn:RemoveClientIDFromOpenIDConnectProvider', removeClientIDFromOpenIDConnectProvider_openIDConnectProviderArn - The Amazon Resource Name (ARN) of the IAM OIDC provider resource to remove the client ID from. You can get a list of OIDC provider ARNs by using the ListOpenIDConnectProviders operation.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

$sel:clientID:RemoveClientIDFromOpenIDConnectProvider', removeClientIDFromOpenIDConnectProvider_clientID - The client ID (also known as audience) to remove from the IAM OIDC provider resource. For more information about client IDs, see CreateOpenIDConnectProvider.

Request Lenses

removeClientIDFromOpenIDConnectProvider_openIDConnectProviderArn :: Lens' RemoveClientIDFromOpenIDConnectProvider Text

The Amazon Resource Name (ARN) of the IAM OIDC provider resource to remove the client ID from. You can get a list of OIDC provider ARNs by using the ListOpenIDConnectProviders operation.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

removeClientIDFromOpenIDConnectProvider_clientID :: Lens' RemoveClientIDFromOpenIDConnectProvider Text

The client ID (also known as audience) to remove from the IAM OIDC provider resource. For more information about client IDs, see CreateOpenIDConnectProvider.

Destructuring the Response

data RemoveClientIDFromOpenIDConnectProviderResponse

See: newRemoveClientIDFromOpenIDConnectProviderResponse smart constructor.

Constructors

RemoveClientIDFromOpenIDConnectProviderResponse'

newRemoveClientIDFromOpenIDConnectProviderResponse :: RemoveClientIDFromOpenIDConnectProviderResponse

Create a value of RemoveClientIDFromOpenIDConnectProviderResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.