Network.AWS.IAM.RemoveRoleFromInstanceProfile

Description

Removes the specified IAM role from the specified EC2 instance profile.

Make sure that you do not have any Amazon EC2 instances running with the role you are about to remove from the instance profile. Removing a role from an instance profile that is associated with a running instance might break any applications running on the instance.

For more information about IAM roles, see Working with roles. For more information about instance profiles, see About instance profiles.

Creating a Request

data RemoveRoleFromInstanceProfile

See: newRemoveRoleFromInstanceProfile smart constructor.

Constructors

RemoveRoleFromInstanceProfile'

Fields

instanceProfileName :: Text

The name of the instance profile to update.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

roleName :: Text

The name of the role to remove.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

newRemoveRoleFromInstanceProfile

Arguments

:: Text
$sel:instanceProfileName:RemoveRoleFromInstanceProfile'

-> Text
$sel:roleName:RemoveRoleFromInstanceProfile'

-> RemoveRoleFromInstanceProfile

Create a value of RemoveRoleFromInstanceProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceProfileName:RemoveRoleFromInstanceProfile', removeRoleFromInstanceProfile_instanceProfileName - The name of the instance profile to update.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:roleName:RemoveRoleFromInstanceProfile', removeRoleFromInstanceProfile_roleName - The name of the role to remove.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Request Lenses

removeRoleFromInstanceProfile_instanceProfileName :: Lens' RemoveRoleFromInstanceProfile Text

The name of the instance profile to update.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

removeRoleFromInstanceProfile_roleName :: Lens' RemoveRoleFromInstanceProfile Text

The name of the role to remove.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Destructuring the Response

data RemoveRoleFromInstanceProfileResponse

See: newRemoveRoleFromInstanceProfileResponse smart constructor.

Constructors

RemoveRoleFromInstanceProfileResponse'

newRemoveRoleFromInstanceProfileResponse :: RemoveRoleFromInstanceProfileResponse

Create a value of RemoveRoleFromInstanceProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.