Network.AWS.IAM.ResyncMFADevice

Description

Synchronizes the specified MFA device with its IAM resource object on the Amazon Web Services servers.

For more information about creating and working with virtual MFA devices, see Using a virtual MFA device in the IAM User Guide.

Creating a Request

data ResyncMFADevice

See: newResyncMFADevice smart constructor.

Constructors

ResyncMFADevice'

Fields

userName :: Text

The name of the user whose MFA device you want to resynchronize.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

serialNumber :: Text

Serial number that uniquely identifies the MFA device.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

authenticationCode1 :: Text

An authentication code emitted by the device.

The format for this parameter is a sequence of six digits.

authenticationCode2 :: Text

A subsequent authentication code emitted by the device.

The format for this parameter is a sequence of six digits.

newResyncMFADevice

Arguments

:: Text
$sel:userName:ResyncMFADevice'

-> Text
$sel:serialNumber:ResyncMFADevice'

-> Text
$sel:authenticationCode1:ResyncMFADevice'

-> Text
$sel:authenticationCode2:ResyncMFADevice'

-> ResyncMFADevice

Create a value of ResyncMFADevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userName:ResyncMFADevice', resyncMFADevice_userName - The name of the user whose MFA device you want to resynchronize.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:serialNumber:ResyncMFADevice', resyncMFADevice_serialNumber - Serial number that uniquely identifies the MFA device.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:authenticationCode1:ResyncMFADevice', resyncMFADevice_authenticationCode1 - An authentication code emitted by the device.

The format for this parameter is a sequence of six digits.

$sel:authenticationCode2:ResyncMFADevice', resyncMFADevice_authenticationCode2 - A subsequent authentication code emitted by the device.

The format for this parameter is a sequence of six digits.

Request Lenses

resyncMFADevice_userName :: Lens' ResyncMFADevice Text

The name of the user whose MFA device you want to resynchronize.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

resyncMFADevice_serialNumber :: Lens' ResyncMFADevice Text

Serial number that uniquely identifies the MFA device.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

resyncMFADevice_authenticationCode1 :: Lens' ResyncMFADevice Text

An authentication code emitted by the device.

The format for this parameter is a sequence of six digits.

resyncMFADevice_authenticationCode2 :: Lens' ResyncMFADevice Text

A subsequent authentication code emitted by the device.

The format for this parameter is a sequence of six digits.

Destructuring the Response

data ResyncMFADeviceResponse

See: newResyncMFADeviceResponse smart constructor.

Constructors

ResyncMFADeviceResponse'

newResyncMFADeviceResponse :: ResyncMFADeviceResponse

Create a value of ResyncMFADeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.