Network.AWS.IAM.TagOpenIDConnectProvider

Description

Adds one or more tags to an OpenID Connect (OIDC)-compatible identity provider. For more information about these providers, see About web identity federation. If a tag with the same key name already exists, then that tag is overwritten with the new value.

A tag consists of a key name and an associated value. By assigning tags to your resources, you can do the following:

Administrative grouping and discovery - Attach tags to resources to aid in organization and search. For example, you could search for all resources with the key name Project and the value MyImportantProject. Or search for all resources with the key name Cost Center and the value 41200.

Access control - Include tags in IAM user-based and resource-based policies. You can use tags to restrict access to only an OIDC provider that has a specified tag attached. For examples of policies that show how to use tags to control access, see Control access using IAM tags in the IAM User Guide.

If any one of the tags is invalid or if you exceed the allowed maximum number of tags, then the entire request fails and the resource is not created. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

Amazon Web Services always interprets the tag Value as a single string. If you need to store an array, you can store comma-separated values in the string. However, you must interpret the value in your code.

Creating a Request

data TagOpenIDConnectProvider

See: newTagOpenIDConnectProvider smart constructor.

Constructors

TagOpenIDConnectProvider'

Fields

openIDConnectProviderArn :: Text

The ARN of the OIDC identity provider in IAM to which you want to add tags.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

tags :: [Tag]

The list of tags that you want to attach to the OIDC identity provider in IAM. Each tag consists of a key name and an associated value.

newTagOpenIDConnectProvider

Arguments

:: Text
$sel:openIDConnectProviderArn:TagOpenIDConnectProvider'

-> TagOpenIDConnectProvider

Create a value of TagOpenIDConnectProvider with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:openIDConnectProviderArn:TagOpenIDConnectProvider', tagOpenIDConnectProvider_openIDConnectProviderArn - The ARN of the OIDC identity provider in IAM to which you want to add tags.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:tags:TagOpenIDConnectProvider', tagOpenIDConnectProvider_tags - The list of tags that you want to attach to the OIDC identity provider in IAM. Each tag consists of a key name and an associated value.

Request Lenses

tagOpenIDConnectProvider_openIDConnectProviderArn :: Lens' TagOpenIDConnectProvider Text

The ARN of the OIDC identity provider in IAM to which you want to add tags.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

tagOpenIDConnectProvider_tags :: Lens' TagOpenIDConnectProvider [Tag]

The list of tags that you want to attach to the OIDC identity provider in IAM. Each tag consists of a key name and an associated value.

Destructuring the Response

data TagOpenIDConnectProviderResponse

See: newTagOpenIDConnectProviderResponse smart constructor.

Constructors

TagOpenIDConnectProviderResponse'

newTagOpenIDConnectProviderResponse :: TagOpenIDConnectProviderResponse

Create a value of TagOpenIDConnectProviderResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.