Network.AWS.IAM.TagPolicy

Description

Adds one or more tags to an IAM customer managed policy. If a tag with the same key name already exists, then that tag is overwritten with the new value.

A tag consists of a key name and an associated value. By assigning tags to your resources, you can do the following:

Administrative grouping and discovery - Attach tags to resources to aid in organization and search. For example, you could search for all resources with the key name Project and the value MyImportantProject. Or search for all resources with the key name Cost Center and the value 41200.

Access control - Include tags in IAM user-based and resource-based policies. You can use tags to restrict access to only an IAM customer managed policy that has a specified tag attached. For examples of policies that show how to use tags to control access, see Control access using IAM tags in the IAM User Guide.

If any one of the tags is invalid or if you exceed the allowed maximum number of tags, then the entire request fails and the resource is not created. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

Amazon Web Services always interprets the tag Value as a single string. If you need to store an array, you can store comma-separated values in the string. However, you must interpret the value in your code.

Creating a Request

data TagPolicy

See: newTagPolicy smart constructor.

Constructors

TagPolicy'

Fields

policyArn :: Text

The ARN of the IAM customer managed policy to which you want to add tags.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

tags :: [Tag]

The list of tags that you want to attach to the IAM customer managed policy. Each tag consists of a key name and an associated value.

newTagPolicy

Arguments

:: Text
$sel:policyArn:TagPolicy'

-> TagPolicy

Create a value of TagPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyArn:TagPolicy', tagPolicy_policyArn - The ARN of the IAM customer managed policy to which you want to add tags.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:tags:TagPolicy', tagPolicy_tags - The list of tags that you want to attach to the IAM customer managed policy. Each tag consists of a key name and an associated value.

Request Lenses

tagPolicy_policyArn :: Lens' TagPolicy Text

The ARN of the IAM customer managed policy to which you want to add tags.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

tagPolicy_tags :: Lens' TagPolicy [Tag]

The list of tags that you want to attach to the IAM customer managed policy. Each tag consists of a key name and an associated value.

Destructuring the Response

data TagPolicyResponse

See: newTagPolicyResponse smart constructor.

Constructors

TagPolicyResponse'

newTagPolicyResponse :: TagPolicyResponse

Create a value of TagPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.