Network.AWS.IAM.TagRole

Description

Adds one or more tags to an IAM role. The role can be a regular role or a service-linked role. If a tag with the same key name already exists, then that tag is overwritten with the new value.

A tag consists of a key name and an associated value. By assigning tags to your resources, you can do the following:

Administrative grouping and discovery - Attach tags to resources to aid in organization and search. For example, you could search for all resources with the key name Project and the value MyImportantProject. Or search for all resources with the key name Cost Center and the value 41200.

Access control - Include tags in IAM user-based and resource-based policies. You can use tags to restrict access to only an IAM role that has a specified tag attached. You can also restrict access to only those resources that have a certain tag attached. For examples of policies that show how to use tags to control access, see Control access using IAM tags in the IAM User Guide.

Cost allocation - Use tags to help track which individuals and teams are using which Amazon Web Services resources.

If any one of the tags is invalid or if you exceed the allowed maximum number of tags, then the entire request fails and the resource is not created. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

Amazon Web Services always interprets the tag Value as a single string. If you need to store an array, you can store comma-separated values in the string. However, you must interpret the value in your code.

For more information about tagging, see Tagging IAM identities in the IAM User Guide.

Creating a Request

data TagRole

See: newTagRole smart constructor.

Constructors

TagRole'

Fields

roleName :: Text

The name of the IAM role to which you want to add tags.

This parameter accepts (through its regex pattern) a string of characters that consist of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

tags :: [Tag]

The list of tags that you want to attach to the IAM role. Each tag consists of a key name and an associated value.

newTagRole

Arguments

:: Text
$sel:roleName:TagRole'

-> TagRole

Create a value of TagRole with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleName:TagRole', tagRole_roleName - The name of the IAM role to which you want to add tags.

This parameter accepts (through its regex pattern) a string of characters that consist of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:tags:TagRole', tagRole_tags - The list of tags that you want to attach to the IAM role. Each tag consists of a key name and an associated value.

Request Lenses

tagRole_roleName :: Lens' TagRole Text

The name of the IAM role to which you want to add tags.

This parameter accepts (through its regex pattern) a string of characters that consist of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

tagRole_tags :: Lens' TagRole [Tag]

The list of tags that you want to attach to the IAM role. Each tag consists of a key name and an associated value.

Destructuring the Response

data TagRoleResponse

See: newTagRoleResponse smart constructor.

Constructors

TagRoleResponse'

newTagRoleResponse :: TagRoleResponse

Create a value of TagRoleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.