Network.AWS.IAM.Types.AccessKeyInfo

Description

Documentation

data AccessKeyInfo

Contains information about an Amazon Web Services access key.

This data type is used as a response element in the CreateAccessKey and ListAccessKeys operations.

The SecretAccessKey value is returned only in response to CreateAccessKey. You can get a secret access key only when you first create an access key; you cannot recover the secret access key later. If you lose a secret access key, you must create a new access key.

See: newAccessKeyInfo smart constructor.

Constructors

AccessKeyInfo'

Fields

createDate :: Maybe ISO8601

The date when the access key was created.

userName :: Text

The name of the IAM user that the access key is associated with.

accessKeyId :: AccessKey

The ID for this access key.

status :: StatusType

The status of the access key. Active means that the key is valid for API calls, while Inactive means it is not.

secretAccessKey :: Sensitive Text

The secret key used to sign requests.

newAccessKeyInfo

Arguments

:: Text
$sel:userName:AccessKeyInfo'

-> AccessKey
$sel:accessKeyId:AccessKeyInfo'

-> StatusType
$sel:status:AccessKeyInfo'

-> Text
$sel:secretAccessKey:AccessKeyInfo'

-> AccessKeyInfo

Create a value of AccessKeyInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createDate:AccessKeyInfo', accessKeyInfo_createDate - The date when the access key was created.

$sel:userName:AccessKeyInfo', accessKeyInfo_userName - The name of the IAM user that the access key is associated with.

$sel:accessKeyId:AccessKeyInfo', accessKeyInfo_accessKeyId - The ID for this access key.

$sel:status:AccessKeyInfo', accessKeyInfo_status - The status of the access key. Active means that the key is valid for API calls, while Inactive means it is not.

$sel:secretAccessKey:AccessKeyInfo', accessKeyInfo_secretAccessKey - The secret key used to sign requests.

accessKeyInfo_createDate :: Lens' AccessKeyInfo (Maybe UTCTime)

The date when the access key was created.

accessKeyInfo_userName :: Lens' AccessKeyInfo Text

The name of the IAM user that the access key is associated with.

accessKeyInfo_accessKeyId :: Lens' AccessKeyInfo AccessKey

The ID for this access key.

accessKeyInfo_status :: Lens' AccessKeyInfo StatusType

The status of the access key. Active means that the key is valid for API calls, while Inactive means it is not.

accessKeyInfo_secretAccessKey :: Lens' AccessKeyInfo Text

The secret key used to sign requests.