Network.AWS.IAM.Types.ContextEntry

Description

Documentation

data ContextEntry

Contains information about a condition context key. It includes the name of the key and specifies the value (or values, if the context key supports multiple values) to use in the simulation. This information is used when evaluating the Condition elements of the input policies.

This data type is used as an input parameter to SimulateCustomPolicy and SimulatePrincipalPolicy.

See: newContextEntry smart constructor.

Constructors

ContextEntry'

Fields

contextKeyValues :: Maybe [Text]

The value (or values, if the condition context key supports multiple values) to provide to the simulation when the key is referenced by a Condition element in an input policy.

contextKeyName :: Maybe Text

The full name of a condition context key, including the service prefix. For example, aws:SourceIp or s3:VersionId.

contextKeyType :: Maybe ContextKeyTypeEnum

The data type of the value (or values) specified in the ContextKeyValues parameter.

newContextEntry :: ContextEntry

Create a value of ContextEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contextKeyValues:ContextEntry', contextEntry_contextKeyValues - The value (or values, if the condition context key supports multiple values) to provide to the simulation when the key is referenced by a Condition element in an input policy.

$sel:contextKeyName:ContextEntry', contextEntry_contextKeyName - The full name of a condition context key, including the service prefix. For example, aws:SourceIp or s3:VersionId.

$sel:contextKeyType:ContextEntry', contextEntry_contextKeyType - The data type of the value (or values) specified in the ContextKeyValues parameter.

contextEntry_contextKeyValues :: Lens' ContextEntry (Maybe [Text])

The value (or values, if the condition context key supports multiple values) to provide to the simulation when the key is referenced by a Condition element in an input policy.

contextEntry_contextKeyName :: Lens' ContextEntry (Maybe Text)

The full name of a condition context key, including the service prefix. For example, aws:SourceIp or s3:VersionId.

contextEntry_contextKeyType :: Lens' ContextEntry (Maybe ContextKeyTypeEnum)

The data type of the value (or values) specified in the ContextKeyValues parameter.