Network.AWS.IAM.Types.ErrorDetails

Description

Documentation

data ErrorDetails

Contains information about the reason that the operation failed.

This data type is used as a response element in the GetOrganizationsAccessReport, GetServiceLastAccessedDetails, and GetServiceLastAccessedDetailsWithEntities operations.

See: newErrorDetails smart constructor.

Constructors

ErrorDetails'

Fields

message :: Text

Detailed information about the reason that the operation failed.

code :: Text

The error code associated with the operation failure.

newErrorDetails

Arguments

:: Text
$sel:message:ErrorDetails'

-> Text
$sel:code:ErrorDetails'

-> ErrorDetails

Create a value of ErrorDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:message:ErrorDetails', errorDetails_message - Detailed information about the reason that the operation failed.

$sel:code:ErrorDetails', errorDetails_code - The error code associated with the operation failure.

errorDetails_message :: Lens' ErrorDetails Text

Detailed information about the reason that the operation failed.

errorDetails_code :: Lens' ErrorDetails Text

The error code associated with the operation failure.