Network.AWS.IAM.Types.MFADevice

Description

Documentation

data MFADevice

Contains information about an MFA device.

This data type is used as a response element in the ListMFADevices operation.

See: newMFADevice smart constructor.

Constructors

MFADevice'

Fields

userName :: Text

The user with whom the MFA device is associated.

serialNumber :: Text

The serial number that uniquely identifies the MFA device. For virtual MFA devices, the serial number is the device ARN.

enableDate :: ISO8601

The date when the MFA device was enabled for the user.

newMFADevice

Arguments

:: Text
$sel:userName:MFADevice'

-> Text
$sel:serialNumber:MFADevice'

-> UTCTime
$sel:enableDate:MFADevice'

-> MFADevice

Create a value of MFADevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userName:MFADevice', mfaDevice_userName - The user with whom the MFA device is associated.

$sel:serialNumber:MFADevice', mfaDevice_serialNumber - The serial number that uniquely identifies the MFA device. For virtual MFA devices, the serial number is the device ARN.

$sel:enableDate:MFADevice', mfaDevice_enableDate - The date when the MFA device was enabled for the user.

mfaDevice_userName :: Lens' MFADevice Text

The user with whom the MFA device is associated.

mfaDevice_serialNumber :: Lens' MFADevice Text

The serial number that uniquely identifies the MFA device. For virtual MFA devices, the serial number is the device ARN.

mfaDevice_enableDate :: Lens' MFADevice UTCTime

The date when the MFA device was enabled for the user.