Network.AWS.IAM.Types.ManagedPolicyDetail

Description

Documentation

data ManagedPolicyDetail

Contains information about a managed policy, including the policy's ARN, versions, and the number of principal entities (users, groups, and roles) that the policy is attached to.

This data type is used as a response element in the GetAccountAuthorizationDetails operation.

For more information about managed policies, see Managed policies and inline policies in the IAM User Guide.

See: newManagedPolicyDetail smart constructor.

Constructors

ManagedPolicyDetail'

Fields

policyName :: Maybe Text

The friendly name (not ARN) identifying the policy.

permissionsBoundaryUsageCount :: Maybe Int

The number of entities (users and roles) for which the policy is used as the permissions boundary.

For more information about permissions boundaries, see Permissions boundaries for IAM identities in the IAM User Guide.

isAttachable :: Maybe Bool

Specifies whether the policy can be attached to an IAM user, group, or role.

createDate :: Maybe ISO8601

The date and time, in ISO 8601 date-time format, when the policy was created.

arn :: Maybe Text

attachmentCount :: Maybe Int

The number of principal entities (users, groups, and roles) that the policy is attached to.

defaultVersionId :: Maybe Text

The identifier for the version of the policy that is set as the default (operative) version.

For more information about policy versions, see Versioning for managed policies in the IAM User Guide.

description :: Maybe Text

A friendly description of the policy.

policyVersionList :: Maybe [PolicyVersion]

A list containing information about the versions of the policy.

updateDate :: Maybe ISO8601

The date and time, in ISO 8601 date-time format, when the policy was last updated.

When a policy has only one version, this field contains the date and time when the policy was created. When a policy has more than one version, this field contains the date and time when the most recent policy version was created.

path :: Maybe Text

The path to the policy.

For more information about paths, see IAM identifiers in the IAM User Guide.

policyId :: Maybe Text

The stable and unique string identifying the policy.

For more information about IDs, see IAM identifiers in the IAM User Guide.

newManagedPolicyDetail :: ManagedPolicyDetail

Create a value of ManagedPolicyDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyName:ManagedPolicyDetail', managedPolicyDetail_policyName - The friendly name (not ARN) identifying the policy.

$sel:permissionsBoundaryUsageCount:ManagedPolicyDetail', managedPolicyDetail_permissionsBoundaryUsageCount - The number of entities (users and roles) for which the policy is used as the permissions boundary.

For more information about permissions boundaries, see Permissions boundaries for IAM identities in the IAM User Guide.

$sel:isAttachable:ManagedPolicyDetail', managedPolicyDetail_isAttachable - Specifies whether the policy can be attached to an IAM user, group, or role.

$sel:createDate:ManagedPolicyDetail', managedPolicyDetail_createDate - The date and time, in ISO 8601 date-time format, when the policy was created.

$sel:arn:ManagedPolicyDetail', managedPolicyDetail_arn - Undocumented member.

$sel:attachmentCount:ManagedPolicyDetail', managedPolicyDetail_attachmentCount - The number of principal entities (users, groups, and roles) that the policy is attached to.

$sel:defaultVersionId:ManagedPolicyDetail', managedPolicyDetail_defaultVersionId - The identifier for the version of the policy that is set as the default (operative) version.

For more information about policy versions, see Versioning for managed policies in the IAM User Guide.

$sel:description:ManagedPolicyDetail', managedPolicyDetail_description - A friendly description of the policy.

$sel:policyVersionList:ManagedPolicyDetail', managedPolicyDetail_policyVersionList - A list containing information about the versions of the policy.

$sel:updateDate:ManagedPolicyDetail', managedPolicyDetail_updateDate - The date and time, in ISO 8601 date-time format, when the policy was last updated.

When a policy has only one version, this field contains the date and time when the policy was created. When a policy has more than one version, this field contains the date and time when the most recent policy version was created.

$sel:path:ManagedPolicyDetail', managedPolicyDetail_path - The path to the policy.

For more information about paths, see IAM identifiers in the IAM User Guide.

$sel:policyId:ManagedPolicyDetail', managedPolicyDetail_policyId - The stable and unique string identifying the policy.

For more information about IDs, see IAM identifiers in the IAM User Guide.

managedPolicyDetail_policyName :: Lens' ManagedPolicyDetail (Maybe Text)

The friendly name (not ARN) identifying the policy.

managedPolicyDetail_permissionsBoundaryUsageCount :: Lens' ManagedPolicyDetail (Maybe Int)

The number of entities (users and roles) for which the policy is used as the permissions boundary.

For more information about permissions boundaries, see Permissions boundaries for IAM identities in the IAM User Guide.

managedPolicyDetail_isAttachable :: Lens' ManagedPolicyDetail (Maybe Bool)

Specifies whether the policy can be attached to an IAM user, group, or role.

managedPolicyDetail_createDate :: Lens' ManagedPolicyDetail (Maybe UTCTime)

The date and time, in ISO 8601 date-time format, when the policy was created.

managedPolicyDetail_arn :: Lens' ManagedPolicyDetail (Maybe Text)

Undocumented member.

managedPolicyDetail_attachmentCount :: Lens' ManagedPolicyDetail (Maybe Int)

The number of principal entities (users, groups, and roles) that the policy is attached to.

managedPolicyDetail_defaultVersionId :: Lens' ManagedPolicyDetail (Maybe Text)

The identifier for the version of the policy that is set as the default (operative) version.

For more information about policy versions, see Versioning for managed policies in the IAM User Guide.

managedPolicyDetail_description :: Lens' ManagedPolicyDetail (Maybe Text)

A friendly description of the policy.

managedPolicyDetail_policyVersionList :: Lens' ManagedPolicyDetail (Maybe [PolicyVersion])

A list containing information about the versions of the policy.

managedPolicyDetail_updateDate :: Lens' ManagedPolicyDetail (Maybe UTCTime)

The date and time, in ISO 8601 date-time format, when the policy was last updated.

When a policy has only one version, this field contains the date and time when the policy was created. When a policy has more than one version, this field contains the date and time when the most recent policy version was created.

managedPolicyDetail_path :: Lens' ManagedPolicyDetail (Maybe Text)

The path to the policy.

For more information about paths, see IAM identifiers in the IAM User Guide.

managedPolicyDetail_policyId :: Lens' ManagedPolicyDetail (Maybe Text)

The stable and unique string identifying the policy.

For more information about IDs, see IAM identifiers in the IAM User Guide.