Network.AWS.IAM.Types.PasswordPolicy

Description

Documentation

data PasswordPolicy

Contains information about the account password policy.

This data type is used as a response element in the GetAccountPasswordPolicy operation.

See: newPasswordPolicy smart constructor.

Constructors

PasswordPolicy'

Fields

requireLowercaseCharacters :: Maybe Bool

Specifies whether IAM user passwords must contain at least one lowercase character (a to z).

maxPasswordAge :: Maybe Natural

The number of days that an IAM user password is valid.

passwordReusePrevention :: Maybe Natural

Specifies the number of previous passwords that IAM users are prevented from reusing.

minimumPasswordLength :: Maybe Natural

Minimum length to require for IAM user passwords.

expirePasswords :: Maybe Bool

Indicates whether passwords in the account expire. Returns true if MaxPasswordAge contains a value greater than 0. Returns false if MaxPasswordAge is 0 or not present.

requireUppercaseCharacters :: Maybe Bool

Specifies whether IAM user passwords must contain at least one uppercase character (A to Z).

allowUsersToChangePassword :: Maybe Bool

Specifies whether IAM users are allowed to change their own password.

hardExpiry :: Maybe Bool

Specifies whether IAM users are prevented from setting a new password after their password has expired.

requireSymbols :: Maybe Bool

Specifies whether IAM user passwords must contain at least one of the following symbols:

! @ # $ % ^ & * ( ) _ + - = [ ] { } | '

requireNumbers :: Maybe Bool

Specifies whether IAM user passwords must contain at least one numeric character (0 to 9).

newPasswordPolicy :: PasswordPolicy

Create a value of PasswordPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requireLowercaseCharacters:PasswordPolicy', passwordPolicy_requireLowercaseCharacters - Specifies whether IAM user passwords must contain at least one lowercase character (a to z).

$sel:maxPasswordAge:PasswordPolicy', passwordPolicy_maxPasswordAge - The number of days that an IAM user password is valid.

$sel:passwordReusePrevention:PasswordPolicy', passwordPolicy_passwordReusePrevention - Specifies the number of previous passwords that IAM users are prevented from reusing.

$sel:minimumPasswordLength:PasswordPolicy', passwordPolicy_minimumPasswordLength - Minimum length to require for IAM user passwords.

$sel:expirePasswords:PasswordPolicy', passwordPolicy_expirePasswords - Indicates whether passwords in the account expire. Returns true if MaxPasswordAge contains a value greater than 0. Returns false if MaxPasswordAge is 0 or not present.

$sel:requireUppercaseCharacters:PasswordPolicy', passwordPolicy_requireUppercaseCharacters - Specifies whether IAM user passwords must contain at least one uppercase character (A to Z).

$sel:allowUsersToChangePassword:PasswordPolicy', passwordPolicy_allowUsersToChangePassword - Specifies whether IAM users are allowed to change their own password.

$sel:hardExpiry:PasswordPolicy', passwordPolicy_hardExpiry - Specifies whether IAM users are prevented from setting a new password after their password has expired.

$sel:requireSymbols:PasswordPolicy', passwordPolicy_requireSymbols - Specifies whether IAM user passwords must contain at least one of the following symbols:

! @ # $ % ^ & * ( ) _ + - = [ ] { } | '

$sel:requireNumbers:PasswordPolicy', passwordPolicy_requireNumbers - Specifies whether IAM user passwords must contain at least one numeric character (0 to 9).

passwordPolicy_requireLowercaseCharacters :: Lens' PasswordPolicy (Maybe Bool)

Specifies whether IAM user passwords must contain at least one lowercase character (a to z).

passwordPolicy_maxPasswordAge :: Lens' PasswordPolicy (Maybe Natural)

The number of days that an IAM user password is valid.

passwordPolicy_passwordReusePrevention :: Lens' PasswordPolicy (Maybe Natural)

Specifies the number of previous passwords that IAM users are prevented from reusing.

passwordPolicy_minimumPasswordLength :: Lens' PasswordPolicy (Maybe Natural)

Minimum length to require for IAM user passwords.

passwordPolicy_expirePasswords :: Lens' PasswordPolicy (Maybe Bool)

Indicates whether passwords in the account expire. Returns true if MaxPasswordAge contains a value greater than 0. Returns false if MaxPasswordAge is 0 or not present.

passwordPolicy_requireUppercaseCharacters :: Lens' PasswordPolicy (Maybe Bool)

Specifies whether IAM user passwords must contain at least one uppercase character (A to Z).

passwordPolicy_allowUsersToChangePassword :: Lens' PasswordPolicy (Maybe Bool)

Specifies whether IAM users are allowed to change their own password.

passwordPolicy_hardExpiry :: Lens' PasswordPolicy (Maybe Bool)

Specifies whether IAM users are prevented from setting a new password after their password has expired.

passwordPolicy_requireSymbols :: Lens' PasswordPolicy (Maybe Bool)

Specifies whether IAM user passwords must contain at least one of the following symbols:

! @ # $ % ^ & * ( ) _ + - = [ ] { } | '

passwordPolicy_requireNumbers :: Lens' PasswordPolicy (Maybe Bool)

Specifies whether IAM user passwords must contain at least one numeric character (0 to 9).