Network.AWS.IAM.Types.PolicyDetail

Description

Documentation

data PolicyDetail

Contains information about an IAM policy, including the policy document.

This data type is used as a response element in the GetAccountAuthorizationDetails operation.

See: newPolicyDetail smart constructor.

Constructors

PolicyDetail'

Fields

policyDocument :: Maybe Text

The policy document.

policyName :: Maybe Text

The name of the policy.

newPolicyDetail :: PolicyDetail

Create a value of PolicyDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policyDocument:PolicyDetail', policyDetail_policyDocument - The policy document.

$sel:policyName:PolicyDetail', policyDetail_policyName - The name of the policy.

policyDetail_policyDocument :: Lens' PolicyDetail (Maybe Text)

The policy document.

policyDetail_policyName :: Lens' PolicyDetail (Maybe Text)

The name of the policy.