Network.AWS.IAM.Types.Role

Description

Documentation

data Role

Contains information about an IAM role. This structure is returned as a response element in several API operations that interact with roles.

See: newRole smart constructor.

Constructors

Role'

Fields

maxSessionDuration :: Maybe Natural

The maximum session duration (in seconds) for the specified role. Anyone who uses the CLI, or API to assume the role can specify the duration using the optional DurationSeconds API parameter or duration-seconds CLI parameter.

assumeRolePolicyDocument :: Maybe Text

The policy that grants an entity permission to assume the role.

roleLastUsed :: Maybe RoleLastUsed

Contains information about the last time that an IAM role was used. This includes the date and time and the Region in which the role was last used. Activity is only reported for the trailing 400 days. This period can be shorter if your Region began supporting these features within the last year. The role might have been used more than 400 days ago. For more information, see Regions where data is tracked in the IAM User Guide.

permissionsBoundary :: Maybe AttachedPermissionsBoundary

The ARN of the policy used to set the permissions boundary for the role.

For more information about permissions boundaries, see Permissions boundaries for IAM identities in the IAM User Guide.

description :: Maybe Text

A description of the role that you provide.

tags :: Maybe [Tag]

A list of tags that are attached to the role. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

path :: Text

The path to the role. For more information about paths, see IAM identifiers in the IAM User Guide.

roleName :: Text

The friendly name that identifies the role.

roleId :: Text

The stable and unique string identifying the role. For more information about IDs, see IAM identifiers in the IAM User Guide.

arn :: Text

The Amazon Resource Name (ARN) specifying the role. For more information about ARNs and how to use them in policies, see IAM identifiers in the IAM User Guide guide.

createDate :: ISO8601

The date and time, in ISO 8601 date-time format, when the role was created.

newRole

Arguments

:: Text
$sel:path:Role'

-> Text
$sel:roleName:Role'

-> Text
$sel:roleId:Role'

-> Text
$sel:arn:Role'

-> UTCTime
$sel:createDate:Role'

-> Role

Create a value of Role with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxSessionDuration:Role', role_maxSessionDuration - The maximum session duration (in seconds) for the specified role. Anyone who uses the CLI, or API to assume the role can specify the duration using the optional DurationSeconds API parameter or duration-seconds CLI parameter.

$sel:assumeRolePolicyDocument:Role', role_assumeRolePolicyDocument - The policy that grants an entity permission to assume the role.

$sel:roleLastUsed:Role', role_roleLastUsed - Contains information about the last time that an IAM role was used. This includes the date and time and the Region in which the role was last used. Activity is only reported for the trailing 400 days. This period can be shorter if your Region began supporting these features within the last year. The role might have been used more than 400 days ago. For more information, see Regions where data is tracked in the IAM User Guide.

$sel:permissionsBoundary:Role', role_permissionsBoundary - The ARN of the policy used to set the permissions boundary for the role.

For more information about permissions boundaries, see Permissions boundaries for IAM identities in the IAM User Guide.

$sel:description:Role', role_description - A description of the role that you provide.

$sel:tags:Role', role_tags - A list of tags that are attached to the role. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

$sel:path:Role', role_path - The path to the role. For more information about paths, see IAM identifiers in the IAM User Guide.

$sel:roleName:Role', role_roleName - The friendly name that identifies the role.

$sel:roleId:Role', role_roleId - The stable and unique string identifying the role. For more information about IDs, see IAM identifiers in the IAM User Guide.

$sel:arn:Role', role_arn - The Amazon Resource Name (ARN) specifying the role. For more information about ARNs and how to use them in policies, see IAM identifiers in the IAM User Guide guide.

$sel:createDate:Role', role_createDate - The date and time, in ISO 8601 date-time format, when the role was created.

role_maxSessionDuration :: Lens' Role (Maybe Natural)

The maximum session duration (in seconds) for the specified role. Anyone who uses the CLI, or API to assume the role can specify the duration using the optional DurationSeconds API parameter or duration-seconds CLI parameter.

role_assumeRolePolicyDocument :: Lens' Role (Maybe Text)

The policy that grants an entity permission to assume the role.

role_roleLastUsed :: Lens' Role (Maybe RoleLastUsed)

Contains information about the last time that an IAM role was used. This includes the date and time and the Region in which the role was last used. Activity is only reported for the trailing 400 days. This period can be shorter if your Region began supporting these features within the last year. The role might have been used more than 400 days ago. For more information, see Regions where data is tracked in the IAM User Guide.

role_permissionsBoundary :: Lens' Role (Maybe AttachedPermissionsBoundary)

The ARN of the policy used to set the permissions boundary for the role.

For more information about permissions boundaries, see Permissions boundaries for IAM identities in the IAM User Guide.

role_description :: Lens' Role (Maybe Text)

A description of the role that you provide.

role_tags :: Lens' Role (Maybe [Tag])

A list of tags that are attached to the role. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

role_path :: Lens' Role Text

The path to the role. For more information about paths, see IAM identifiers in the IAM User Guide.

role_roleName :: Lens' Role Text

The friendly name that identifies the role.

role_roleId :: Lens' Role Text

The stable and unique string identifying the role. For more information about IDs, see IAM identifiers in the IAM User Guide.

role_arn :: Lens' Role Text

The Amazon Resource Name (ARN) specifying the role. For more information about ARNs and how to use them in policies, see IAM identifiers in the IAM User Guide guide.

role_createDate :: Lens' Role UTCTime

The date and time, in ISO 8601 date-time format, when the role was created.