Network.AWS.IAM.Types.RoleDetail

Description

Documentation

data RoleDetail

Contains information about an IAM role, including all of the role's policies.

This data type is used as a response element in the GetAccountAuthorizationDetails operation.

See: newRoleDetail smart constructor.

Constructors

RoleDetail'

Fields

assumeRolePolicyDocument :: Maybe Text

The trust policy that grants permission to assume the role.

roleLastUsed :: Maybe RoleLastUsed

Contains information about the last time that an IAM role was used. This includes the date and time and the Region in which the role was last used. Activity is only reported for the trailing 400 days. This period can be shorter if your Region began supporting these features within the last year. The role might have been used more than 400 days ago. For more information, see Regions where data is tracked in the IAM User Guide.

roleId :: Maybe Text

The stable and unique string identifying the role. For more information about IDs, see IAM identifiers in the IAM User Guide.

attachedManagedPolicies :: Maybe [AttachedPolicy]

A list of managed policies attached to the role. These policies are the role's access (permissions) policies.

rolePolicyList :: Maybe [PolicyDetail]

A list of inline policies embedded in the role. These policies are the role's access (permissions) policies.

roleName :: Maybe Text

The friendly name that identifies the role.

createDate :: Maybe ISO8601

The date and time, in ISO 8601 date-time format, when the role was created.

permissionsBoundary :: Maybe AttachedPermissionsBoundary

The ARN of the policy used to set the permissions boundary for the role.

For more information about permissions boundaries, see Permissions boundaries for IAM identities in the IAM User Guide.

arn :: Maybe Text

tags :: Maybe [Tag]

A list of tags that are attached to the role. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

instanceProfileList :: Maybe [InstanceProfile]

A list of instance profiles that contain this role.

path :: Maybe Text

The path to the role. For more information about paths, see IAM identifiers in the IAM User Guide.

newRoleDetail :: RoleDetail

Create a value of RoleDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assumeRolePolicyDocument:RoleDetail', roleDetail_assumeRolePolicyDocument - The trust policy that grants permission to assume the role.

$sel:roleLastUsed:RoleDetail', roleDetail_roleLastUsed - Contains information about the last time that an IAM role was used. This includes the date and time and the Region in which the role was last used. Activity is only reported for the trailing 400 days. This period can be shorter if your Region began supporting these features within the last year. The role might have been used more than 400 days ago. For more information, see Regions where data is tracked in the IAM User Guide.

$sel:roleId:RoleDetail', roleDetail_roleId - The stable and unique string identifying the role. For more information about IDs, see IAM identifiers in the IAM User Guide.

$sel:attachedManagedPolicies:RoleDetail', roleDetail_attachedManagedPolicies - A list of managed policies attached to the role. These policies are the role's access (permissions) policies.

$sel:rolePolicyList:RoleDetail', roleDetail_rolePolicyList - A list of inline policies embedded in the role. These policies are the role's access (permissions) policies.

$sel:roleName:RoleDetail', roleDetail_roleName - The friendly name that identifies the role.

$sel:createDate:RoleDetail', roleDetail_createDate - The date and time, in ISO 8601 date-time format, when the role was created.

$sel:permissionsBoundary:RoleDetail', roleDetail_permissionsBoundary - The ARN of the policy used to set the permissions boundary for the role.

For more information about permissions boundaries, see Permissions boundaries for IAM identities in the IAM User Guide.

$sel:arn:RoleDetail', roleDetail_arn - Undocumented member.

$sel:tags:RoleDetail', roleDetail_tags - A list of tags that are attached to the role. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

$sel:instanceProfileList:RoleDetail', roleDetail_instanceProfileList - A list of instance profiles that contain this role.

$sel:path:RoleDetail', roleDetail_path - The path to the role. For more information about paths, see IAM identifiers in the IAM User Guide.

roleDetail_assumeRolePolicyDocument :: Lens' RoleDetail (Maybe Text)

The trust policy that grants permission to assume the role.

roleDetail_roleLastUsed :: Lens' RoleDetail (Maybe RoleLastUsed)

Contains information about the last time that an IAM role was used. This includes the date and time and the Region in which the role was last used. Activity is only reported for the trailing 400 days. This period can be shorter if your Region began supporting these features within the last year. The role might have been used more than 400 days ago. For more information, see Regions where data is tracked in the IAM User Guide.

roleDetail_roleId :: Lens' RoleDetail (Maybe Text)

The stable and unique string identifying the role. For more information about IDs, see IAM identifiers in the IAM User Guide.

roleDetail_attachedManagedPolicies :: Lens' RoleDetail (Maybe [AttachedPolicy])

A list of managed policies attached to the role. These policies are the role's access (permissions) policies.

roleDetail_rolePolicyList :: Lens' RoleDetail (Maybe [PolicyDetail])

A list of inline policies embedded in the role. These policies are the role's access (permissions) policies.

roleDetail_roleName :: Lens' RoleDetail (Maybe Text)

The friendly name that identifies the role.

roleDetail_createDate :: Lens' RoleDetail (Maybe UTCTime)

The date and time, in ISO 8601 date-time format, when the role was created.

roleDetail_permissionsBoundary :: Lens' RoleDetail (Maybe AttachedPermissionsBoundary)

The ARN of the policy used to set the permissions boundary for the role.

For more information about permissions boundaries, see Permissions boundaries for IAM identities in the IAM User Guide.

roleDetail_arn :: Lens' RoleDetail (Maybe Text)

Undocumented member.

roleDetail_tags :: Lens' RoleDetail (Maybe [Tag])

A list of tags that are attached to the role. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

roleDetail_instanceProfileList :: Lens' RoleDetail (Maybe [InstanceProfile])

A list of instance profiles that contain this role.

roleDetail_path :: Lens' RoleDetail (Maybe Text)

The path to the role. For more information about paths, see IAM identifiers in the IAM User Guide.