Network.AWS.IAM.Types.RoleUsageType

Description

Documentation

data RoleUsageType

An object that contains details about how a service-linked role is used, if that information is returned by the service.

This data type is used as a response element in the GetServiceLinkedRoleDeletionStatus operation.

See: newRoleUsageType smart constructor.

Constructors

RoleUsageType'

Fields

resources :: Maybe [Text]

The name of the resource that is using the service-linked role.

region :: Maybe Text

The name of the Region where the service-linked role is being used.

newRoleUsageType :: RoleUsageType

Create a value of RoleUsageType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resources:RoleUsageType', roleUsageType_resources - The name of the resource that is using the service-linked role.

$sel:region:RoleUsageType', roleUsageType_region - The name of the Region where the service-linked role is being used.

roleUsageType_resources :: Lens' RoleUsageType (Maybe [Text])

The name of the resource that is using the service-linked role.

roleUsageType_region :: Lens' RoleUsageType (Maybe Text)

The name of the Region where the service-linked role is being used.