Network.AWS.IAM.Types.ServerCertificate

Description

Documentation

data ServerCertificate

Contains information about a server certificate.

This data type is used as a response element in the GetServerCertificate operation.

See: newServerCertificate smart constructor.

Constructors

ServerCertificate'

Fields

tags :: Maybe [Tag]

A list of tags that are attached to the server certificate. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

certificateChain :: Maybe Text

The contents of the public key certificate chain.

serverCertificateMetadata :: ServerCertificateMetadata

The meta information of the server certificate, such as its name, path, ID, and ARN.

certificateBody :: Text

The contents of the public key certificate.

newServerCertificate

Arguments

:: ServerCertificateMetadata
$sel:serverCertificateMetadata:ServerCertificate'

-> Text
$sel:certificateBody:ServerCertificate'

-> ServerCertificate

Create a value of ServerCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ServerCertificate', serverCertificate_tags - A list of tags that are attached to the server certificate. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

$sel:certificateChain:ServerCertificate', serverCertificate_certificateChain - The contents of the public key certificate chain.

$sel:serverCertificateMetadata:ServerCertificate', serverCertificate_serverCertificateMetadata - The meta information of the server certificate, such as its name, path, ID, and ARN.

$sel:certificateBody:ServerCertificate', serverCertificate_certificateBody - The contents of the public key certificate.

serverCertificate_tags :: Lens' ServerCertificate (Maybe [Tag])

A list of tags that are attached to the server certificate. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

serverCertificate_certificateChain :: Lens' ServerCertificate (Maybe Text)

The contents of the public key certificate chain.

serverCertificate_serverCertificateMetadata :: Lens' ServerCertificate ServerCertificateMetadata

The meta information of the server certificate, such as its name, path, ID, and ARN.

serverCertificate_certificateBody :: Lens' ServerCertificate Text

The contents of the public key certificate.