Network.AWS.IAM.Types.ServerCertificateMetadata

Description

Documentation

data ServerCertificateMetadata

Contains information about a server certificate without its certificate body, certificate chain, and private key.

This data type is used as a response element in the UploadServerCertificate and ListServerCertificates operations.

See: newServerCertificateMetadata smart constructor.

Constructors

ServerCertificateMetadata'

Fields

uploadDate :: Maybe ISO8601

The date when the server certificate was uploaded.

expiration :: Maybe ISO8601

The date on which the certificate is set to expire.

path :: Text

The path to the server certificate. For more information about paths, see IAM identifiers in the IAM User Guide.

serverCertificateName :: Text

The name that identifies the server certificate.

serverCertificateId :: Text

The stable and unique string identifying the server certificate. For more information about IDs, see IAM identifiers in the IAM User Guide.

arn :: Text

The Amazon Resource Name (ARN) specifying the server certificate. For more information about ARNs and how to use them in policies, see IAM identifiers in the IAM User Guide.

newServerCertificateMetadata

Arguments

:: Text
$sel:path:ServerCertificateMetadata'

-> Text
$sel:serverCertificateName:ServerCertificateMetadata'

-> Text
$sel:serverCertificateId:ServerCertificateMetadata'

-> Text
$sel:arn:ServerCertificateMetadata'

-> ServerCertificateMetadata

Create a value of ServerCertificateMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:uploadDate:ServerCertificateMetadata', serverCertificateMetadata_uploadDate - The date when the server certificate was uploaded.

$sel:expiration:ServerCertificateMetadata', serverCertificateMetadata_expiration - The date on which the certificate is set to expire.

$sel:path:ServerCertificateMetadata', serverCertificateMetadata_path - The path to the server certificate. For more information about paths, see IAM identifiers in the IAM User Guide.

$sel:serverCertificateName:ServerCertificateMetadata', serverCertificateMetadata_serverCertificateName - The name that identifies the server certificate.

$sel:serverCertificateId:ServerCertificateMetadata', serverCertificateMetadata_serverCertificateId - The stable and unique string identifying the server certificate. For more information about IDs, see IAM identifiers in the IAM User Guide.

$sel:arn:ServerCertificateMetadata', serverCertificateMetadata_arn - The Amazon Resource Name (ARN) specifying the server certificate. For more information about ARNs and how to use them in policies, see IAM identifiers in the IAM User Guide.

serverCertificateMetadata_uploadDate :: Lens' ServerCertificateMetadata (Maybe UTCTime)

The date when the server certificate was uploaded.

serverCertificateMetadata_expiration :: Lens' ServerCertificateMetadata (Maybe UTCTime)

The date on which the certificate is set to expire.

serverCertificateMetadata_path :: Lens' ServerCertificateMetadata Text

The path to the server certificate. For more information about paths, see IAM identifiers in the IAM User Guide.

serverCertificateMetadata_serverCertificateName :: Lens' ServerCertificateMetadata Text

The name that identifies the server certificate.

serverCertificateMetadata_serverCertificateId :: Lens' ServerCertificateMetadata Text

The stable and unique string identifying the server certificate. For more information about IDs, see IAM identifiers in the IAM User Guide.

serverCertificateMetadata_arn :: Lens' ServerCertificateMetadata Text

The Amazon Resource Name (ARN) specifying the server certificate. For more information about ARNs and how to use them in policies, see IAM identifiers in the IAM User Guide.