Network.AWS.IAM.Types.ServiceLastAccessed

Description

Documentation

data ServiceLastAccessed

Contains details about the most recent attempt to access the service.

This data type is used as a response element in the GetServiceLastAccessedDetails operation.

See: newServiceLastAccessed smart constructor.

Constructors

ServiceLastAccessed'

Fields

totalAuthenticatedEntities :: Maybe Int

The total number of authenticated principals (root user, IAM users, or IAM roles) that have attempted to access the service.

This field is null if no principals attempted to access the service within the reporting period.

lastAuthenticatedRegion :: Maybe Text

The Region from which the authenticated entity (user or role) last attempted to access the service. Amazon Web Services does not report unauthenticated requests.

This field is null if no IAM entities attempted to access the service within the reporting period.

lastAuthenticatedEntity :: Maybe Text

The ARN of the authenticated entity (user or role) that last attempted to access the service. Amazon Web Services does not report unauthenticated requests.

This field is null if no IAM entities attempted to access the service within the reporting period.

trackedActionsLastAccessed :: Maybe [TrackedActionLastAccessed]

An object that contains details about the most recent attempt to access a tracked action within the service.

This field is null if there no tracked actions or if the principal did not use the tracked actions within the reporting period. This field is also null if the report was generated at the service level and not the action level. For more information, see the Granularity field in GenerateServiceLastAccessedDetails.

lastAuthenticated :: Maybe ISO8601

The date and time, in ISO 8601 date-time format, when an authenticated entity most recently attempted to access the service. Amazon Web Services does not report unauthenticated requests.

This field is null if no IAM entities attempted to access the service within the reporting period.

serviceName :: Text

The name of the service in which access was attempted.

serviceNamespace :: Text

The namespace of the service in which access was attempted.

To learn the service namespace of a service, see Actions, resources, and condition keys for Amazon Web Services services in the Service Authorization Reference. Choose the name of the service to view details for that service. In the first paragraph, find the service prefix. For example, (service prefix: a4b). For more information about service namespaces, see Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

newServiceLastAccessed

Arguments

:: Text
$sel:serviceName:ServiceLastAccessed'

-> Text
$sel:serviceNamespace:ServiceLastAccessed'

-> ServiceLastAccessed

Create a value of ServiceLastAccessed with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:totalAuthenticatedEntities:ServiceLastAccessed', serviceLastAccessed_totalAuthenticatedEntities - The total number of authenticated principals (root user, IAM users, or IAM roles) that have attempted to access the service.

This field is null if no principals attempted to access the service within the reporting period.

$sel:lastAuthenticatedRegion:ServiceLastAccessed', serviceLastAccessed_lastAuthenticatedRegion - The Region from which the authenticated entity (user or role) last attempted to access the service. Amazon Web Services does not report unauthenticated requests.

This field is null if no IAM entities attempted to access the service within the reporting period.

$sel:lastAuthenticatedEntity:ServiceLastAccessed', serviceLastAccessed_lastAuthenticatedEntity - The ARN of the authenticated entity (user or role) that last attempted to access the service. Amazon Web Services does not report unauthenticated requests.

This field is null if no IAM entities attempted to access the service within the reporting period.

$sel:trackedActionsLastAccessed:ServiceLastAccessed', serviceLastAccessed_trackedActionsLastAccessed - An object that contains details about the most recent attempt to access a tracked action within the service.

This field is null if there no tracked actions or if the principal did not use the tracked actions within the reporting period. This field is also null if the report was generated at the service level and not the action level. For more information, see the Granularity field in GenerateServiceLastAccessedDetails.

$sel:lastAuthenticated:ServiceLastAccessed', serviceLastAccessed_lastAuthenticated - The date and time, in ISO 8601 date-time format, when an authenticated entity most recently attempted to access the service. Amazon Web Services does not report unauthenticated requests.

This field is null if no IAM entities attempted to access the service within the reporting period.

$sel:serviceName:ServiceLastAccessed', serviceLastAccessed_serviceName - The name of the service in which access was attempted.

$sel:serviceNamespace:ServiceLastAccessed', serviceLastAccessed_serviceNamespace - The namespace of the service in which access was attempted.

To learn the service namespace of a service, see Actions, resources, and condition keys for Amazon Web Services services in the Service Authorization Reference. Choose the name of the service to view details for that service. In the first paragraph, find the service prefix. For example, (service prefix: a4b). For more information about service namespaces, see Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.

serviceLastAccessed_totalAuthenticatedEntities :: Lens' ServiceLastAccessed (Maybe Int)

The total number of authenticated principals (root user, IAM users, or IAM roles) that have attempted to access the service.

This field is null if no principals attempted to access the service within the reporting period.

serviceLastAccessed_lastAuthenticatedRegion :: Lens' ServiceLastAccessed (Maybe Text)

The Region from which the authenticated entity (user or role) last attempted to access the service. Amazon Web Services does not report unauthenticated requests.

This field is null if no IAM entities attempted to access the service within the reporting period.

serviceLastAccessed_lastAuthenticatedEntity :: Lens' ServiceLastAccessed (Maybe Text)

The ARN of the authenticated entity (user or role) that last attempted to access the service. Amazon Web Services does not report unauthenticated requests.

This field is null if no IAM entities attempted to access the service within the reporting period.

serviceLastAccessed_trackedActionsLastAccessed :: Lens' ServiceLastAccessed (Maybe [TrackedActionLastAccessed])

An object that contains details about the most recent attempt to access a tracked action within the service.

This field is null if there no tracked actions or if the principal did not use the tracked actions within the reporting period. This field is also null if the report was generated at the service level and not the action level. For more information, see the Granularity field in GenerateServiceLastAccessedDetails.

serviceLastAccessed_lastAuthenticated :: Lens' ServiceLastAccessed (Maybe UTCTime)

The date and time, in ISO 8601 date-time format, when an authenticated entity most recently attempted to access the service. Amazon Web Services does not report unauthenticated requests.

This field is null if no IAM entities attempted to access the service within the reporting period.

serviceLastAccessed_serviceName :: Lens' ServiceLastAccessed Text

The name of the service in which access was attempted.

serviceLastAccessed_serviceNamespace :: Lens' ServiceLastAccessed Text

The namespace of the service in which access was attempted.

To learn the service namespace of a service, see Actions, resources, and condition keys for Amazon Web Services services in the Service Authorization Reference. Choose the name of the service to view details for that service. In the first paragraph, find the service prefix. For example, (service prefix: a4b). For more information about service namespaces, see Amazon Web Services Service Namespaces in the Amazon Web Services General Reference.