Network.AWS.IAM.Types.ServiceSpecificCredential

Description

Documentation

data ServiceSpecificCredential

Contains the details of a service-specific credential.

See: newServiceSpecificCredential smart constructor.

Constructors

ServiceSpecificCredential'

Fields

createDate :: ISO8601

The date and time, in ISO 8601 date-time format, when the service-specific credential were created.

serviceName :: Text

The name of the service associated with the service-specific credential.

serviceUserName :: Text

The generated user name for the service-specific credential. This value is generated by combining the IAM user's name combined with the ID number of the Amazon Web Services account, as in jane-at-123456789012, for example. This value cannot be configured by the user.

servicePassword :: Sensitive Text

The generated password for the service-specific credential.

serviceSpecificCredentialId :: Text

The unique identifier for the service-specific credential.

userName :: Text

The name of the IAM user associated with the service-specific credential.

status :: StatusType

The status of the service-specific credential. Active means that the key is valid for API calls, while Inactive means it is not.

newServiceSpecificCredential

Arguments

:: UTCTime
$sel:createDate:ServiceSpecificCredential'

-> Text
$sel:serviceName:ServiceSpecificCredential'

-> Text
$sel:serviceUserName:ServiceSpecificCredential'

-> Text
$sel:servicePassword:ServiceSpecificCredential'

-> Text
$sel:serviceSpecificCredentialId:ServiceSpecificCredential'

-> Text
$sel:userName:ServiceSpecificCredential'

-> StatusType
$sel:status:ServiceSpecificCredential'

-> ServiceSpecificCredential

Create a value of ServiceSpecificCredential with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createDate:ServiceSpecificCredential', serviceSpecificCredential_createDate - The date and time, in ISO 8601 date-time format, when the service-specific credential were created.

$sel:serviceName:ServiceSpecificCredential', serviceSpecificCredential_serviceName - The name of the service associated with the service-specific credential.

$sel:serviceUserName:ServiceSpecificCredential', serviceSpecificCredential_serviceUserName - The generated user name for the service-specific credential. This value is generated by combining the IAM user's name combined with the ID number of the Amazon Web Services account, as in jane-at-123456789012, for example. This value cannot be configured by the user.

$sel:servicePassword:ServiceSpecificCredential', serviceSpecificCredential_servicePassword - The generated password for the service-specific credential.

$sel:serviceSpecificCredentialId:ServiceSpecificCredential', serviceSpecificCredential_serviceSpecificCredentialId - The unique identifier for the service-specific credential.

$sel:userName:ServiceSpecificCredential', serviceSpecificCredential_userName - The name of the IAM user associated with the service-specific credential.

$sel:status:ServiceSpecificCredential', serviceSpecificCredential_status - The status of the service-specific credential. Active means that the key is valid for API calls, while Inactive means it is not.

serviceSpecificCredential_createDate :: Lens' ServiceSpecificCredential UTCTime

The date and time, in ISO 8601 date-time format, when the service-specific credential were created.

serviceSpecificCredential_serviceName :: Lens' ServiceSpecificCredential Text

The name of the service associated with the service-specific credential.

serviceSpecificCredential_serviceUserName :: Lens' ServiceSpecificCredential Text

The generated user name for the service-specific credential. This value is generated by combining the IAM user's name combined with the ID number of the Amazon Web Services account, as in jane-at-123456789012, for example. This value cannot be configured by the user.

serviceSpecificCredential_servicePassword :: Lens' ServiceSpecificCredential Text

The generated password for the service-specific credential.

serviceSpecificCredential_serviceSpecificCredentialId :: Lens' ServiceSpecificCredential Text

The unique identifier for the service-specific credential.

serviceSpecificCredential_userName :: Lens' ServiceSpecificCredential Text

The name of the IAM user associated with the service-specific credential.

serviceSpecificCredential_status :: Lens' ServiceSpecificCredential StatusType

The status of the service-specific credential. Active means that the key is valid for API calls, while Inactive means it is not.