Network.AWS.IAM.Types.TrackedActionLastAccessed

Description

Documentation

data TrackedActionLastAccessed

Contains details about the most recent attempt to access an action within the service.

This data type is used as a response element in the GetServiceLastAccessedDetails operation.

See: newTrackedActionLastAccessed smart constructor.

Constructors

TrackedActionLastAccessed'

Fields

lastAccessedTime :: Maybe ISO8601

The date and time, in ISO 8601 date-time format, when an authenticated entity most recently attempted to access the tracked service. Amazon Web Services does not report unauthenticated requests.

This field is null if no IAM entities attempted to access the service within the reporting period.

actionName :: Maybe Text

The name of the tracked action to which access was attempted. Tracked actions are actions that report activity to IAM.

lastAccessedEntity :: Maybe Text

lastAccessedRegion :: Maybe Text

The Region from which the authenticated entity (user or role) last attempted to access the tracked action. Amazon Web Services does not report unauthenticated requests.

This field is null if no IAM entities attempted to access the service within the reporting period.

newTrackedActionLastAccessed :: TrackedActionLastAccessed

Create a value of TrackedActionLastAccessed with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastAccessedTime:TrackedActionLastAccessed', trackedActionLastAccessed_lastAccessedTime - The date and time, in ISO 8601 date-time format, when an authenticated entity most recently attempted to access the tracked service. Amazon Web Services does not report unauthenticated requests.

This field is null if no IAM entities attempted to access the service within the reporting period.

$sel:actionName:TrackedActionLastAccessed', trackedActionLastAccessed_actionName - The name of the tracked action to which access was attempted. Tracked actions are actions that report activity to IAM.

$sel:lastAccessedEntity:TrackedActionLastAccessed', trackedActionLastAccessed_lastAccessedEntity - Undocumented member.

$sel:lastAccessedRegion:TrackedActionLastAccessed', trackedActionLastAccessed_lastAccessedRegion - The Region from which the authenticated entity (user or role) last attempted to access the tracked action. Amazon Web Services does not report unauthenticated requests.

This field is null if no IAM entities attempted to access the service within the reporting period.

trackedActionLastAccessed_lastAccessedTime :: Lens' TrackedActionLastAccessed (Maybe UTCTime)

The date and time, in ISO 8601 date-time format, when an authenticated entity most recently attempted to access the tracked service. Amazon Web Services does not report unauthenticated requests.

This field is null if no IAM entities attempted to access the service within the reporting period.

trackedActionLastAccessed_actionName :: Lens' TrackedActionLastAccessed (Maybe Text)

The name of the tracked action to which access was attempted. Tracked actions are actions that report activity to IAM.

trackedActionLastAccessed_lastAccessedEntity :: Lens' TrackedActionLastAccessed (Maybe Text)

Undocumented member.

trackedActionLastAccessed_lastAccessedRegion :: Lens' TrackedActionLastAccessed (Maybe Text)

The Region from which the authenticated entity (user or role) last attempted to access the tracked action. Amazon Web Services does not report unauthenticated requests.

This field is null if no IAM entities attempted to access the service within the reporting period.