Network.AWS.IAM.Types.User

Description

Documentation

data User

Contains information about an IAM user entity.

This data type is used as a response element in the following operations:

CreateUser

GetUser

ListUsers

See: newUser smart constructor.

Constructors

User'

Fields

passwordLastUsed :: Maybe ISO8601

The date and time, in ISO 8601 date-time format, when the user's password was last used to sign in to an Amazon Web Services website. For a list of Amazon Web Services websites that capture a user's last sign-in time, see the Credential reports topic in the IAM User Guide. If a password is used more than once in a five-minute span, only the first use is returned in this field. If the field is null (no value), then it indicates that they never signed in with a password. This can be because:

The user never had a password.

A password exists but has not been used since IAM started tracking this information on October 20, 2014.

A null value does not mean that the user never had a password. Also, if the user does not currently have a password but had one in the past, then this field contains the date and time the most recent password was used.

This value is returned only in the GetUser and ListUsers operations.

path :: Maybe Text

The path to the user. For more information about paths, see IAM identifiers in the IAM User Guide.

The ARN of the policy used to set the permissions boundary for the user.

permissionsBoundary :: Maybe AttachedPermissionsBoundary

For more information about permissions boundaries, see Permissions boundaries for IAM identities in the IAM User Guide.

tags :: Maybe [Tag]

A list of tags that are associated with the user. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

userName :: Text

The friendly name identifying the user.

userId :: Text

The stable and unique string identifying the user. For more information about IDs, see IAM identifiers in the IAM User Guide.

arn :: Text

The Amazon Resource Name (ARN) that identifies the user. For more information about ARNs and how to use ARNs in policies, see IAM Identifiers in the IAM User Guide.

createDate :: ISO8601

The date and time, in ISO 8601 date-time format, when the user was created.

newUser

Arguments

:: Text
$sel:userName:User'

-> Text
$sel:userId:User'

-> Text
$sel:arn:User'

-> UTCTime
$sel:createDate:User'

-> User

Create a value of User with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:passwordLastUsed:User', user_passwordLastUsed - The date and time, in ISO 8601 date-time format, when the user's password was last used to sign in to an Amazon Web Services website. For a list of Amazon Web Services websites that capture a user's last sign-in time, see the Credential reports topic in the IAM User Guide. If a password is used more than once in a five-minute span, only the first use is returned in this field. If the field is null (no value), then it indicates that they never signed in with a password. This can be because:

The user never had a password.

A password exists but has not been used since IAM started tracking this information on October 20, 2014.

A null value does not mean that the user never had a password. Also, if the user does not currently have a password but had one in the past, then this field contains the date and time the most recent password was used.

This value is returned only in the GetUser and ListUsers operations.

$sel:path:User', user_path - The path to the user. For more information about paths, see IAM identifiers in the IAM User Guide.

The ARN of the policy used to set the permissions boundary for the user.

$sel:permissionsBoundary:User', user_permissionsBoundary - For more information about permissions boundaries, see Permissions boundaries for IAM identities in the IAM User Guide.

$sel:tags:User', user_tags - A list of tags that are associated with the user. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

$sel:userName:User', user_userName - The friendly name identifying the user.

$sel:userId:User', user_userId - The stable and unique string identifying the user. For more information about IDs, see IAM identifiers in the IAM User Guide.

$sel:arn:User', user_arn - The Amazon Resource Name (ARN) that identifies the user. For more information about ARNs and how to use ARNs in policies, see IAM Identifiers in the IAM User Guide.

$sel:createDate:User', user_createDate - The date and time, in ISO 8601 date-time format, when the user was created.

user_passwordLastUsed :: Lens' User (Maybe UTCTime)

The date and time, in ISO 8601 date-time format, when the user's password was last used to sign in to an Amazon Web Services website. For a list of Amazon Web Services websites that capture a user's last sign-in time, see the Credential reports topic in the IAM User Guide. If a password is used more than once in a five-minute span, only the first use is returned in this field. If the field is null (no value), then it indicates that they never signed in with a password. This can be because:

The user never had a password.

A password exists but has not been used since IAM started tracking this information on October 20, 2014.

A null value does not mean that the user never had a password. Also, if the user does not currently have a password but had one in the past, then this field contains the date and time the most recent password was used.

This value is returned only in the GetUser and ListUsers operations.

user_path :: Lens' User (Maybe Text)

The path to the user. For more information about paths, see IAM identifiers in the IAM User Guide.

The ARN of the policy used to set the permissions boundary for the user.

user_permissionsBoundary :: Lens' User (Maybe AttachedPermissionsBoundary)

For more information about permissions boundaries, see Permissions boundaries for IAM identities in the IAM User Guide.

user_tags :: Lens' User (Maybe [Tag])

A list of tags that are associated with the user. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

user_userName :: Lens' User Text

The friendly name identifying the user.

user_userId :: Lens' User Text

The stable and unique string identifying the user. For more information about IDs, see IAM identifiers in the IAM User Guide.

user_arn :: Lens' User Text

The Amazon Resource Name (ARN) that identifies the user. For more information about ARNs and how to use ARNs in policies, see IAM Identifiers in the IAM User Guide.

user_createDate :: Lens' User UTCTime

The date and time, in ISO 8601 date-time format, when the user was created.