Network.AWS.IAM.Types.UserDetail

Description

Documentation

data UserDetail

Contains information about an IAM user, including all the user's policies and all the IAM groups the user is in.

This data type is used as a response element in the GetAccountAuthorizationDetails operation.

See: newUserDetail smart constructor.

Constructors

UserDetail'

Fields

groupList :: Maybe [Text]

A list of IAM groups that the user is in.

arn :: Maybe Text

path :: Maybe Text

The path to the user. For more information about paths, see IAM identifiers in the IAM User Guide.

createDate :: Maybe ISO8601

The date and time, in ISO 8601 date-time format, when the user was created.

userName :: Maybe Text

The friendly name identifying the user.

userId :: Maybe Text

The stable and unique string identifying the user. For more information about IDs, see IAM identifiers in the IAM User Guide.

permissionsBoundary :: Maybe AttachedPermissionsBoundary

The ARN of the policy used to set the permissions boundary for the user.

For more information about permissions boundaries, see Permissions boundaries for IAM identities in the IAM User Guide.

userPolicyList :: Maybe [PolicyDetail]

A list of the inline policies embedded in the user.

tags :: Maybe [Tag]

A list of tags that are associated with the user. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

attachedManagedPolicies :: Maybe [AttachedPolicy]

A list of the managed policies attached to the user.

newUserDetail :: UserDetail

Create a value of UserDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupList:UserDetail', userDetail_groupList - A list of IAM groups that the user is in.

$sel:arn:UserDetail', userDetail_arn - Undocumented member.

$sel:path:UserDetail', userDetail_path - The path to the user. For more information about paths, see IAM identifiers in the IAM User Guide.

$sel:createDate:UserDetail', userDetail_createDate - The date and time, in ISO 8601 date-time format, when the user was created.

$sel:userName:UserDetail', userDetail_userName - The friendly name identifying the user.

$sel:userId:UserDetail', userDetail_userId - The stable and unique string identifying the user. For more information about IDs, see IAM identifiers in the IAM User Guide.

$sel:permissionsBoundary:UserDetail', userDetail_permissionsBoundary - The ARN of the policy used to set the permissions boundary for the user.

For more information about permissions boundaries, see Permissions boundaries for IAM identities in the IAM User Guide.

$sel:userPolicyList:UserDetail', userDetail_userPolicyList - A list of the inline policies embedded in the user.

$sel:tags:UserDetail', userDetail_tags - A list of tags that are associated with the user. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

$sel:attachedManagedPolicies:UserDetail', userDetail_attachedManagedPolicies - A list of the managed policies attached to the user.

userDetail_groupList :: Lens' UserDetail (Maybe [Text])

A list of IAM groups that the user is in.

userDetail_arn :: Lens' UserDetail (Maybe Text)

Undocumented member.

userDetail_path :: Lens' UserDetail (Maybe Text)

The path to the user. For more information about paths, see IAM identifiers in the IAM User Guide.

userDetail_createDate :: Lens' UserDetail (Maybe UTCTime)

The date and time, in ISO 8601 date-time format, when the user was created.

userDetail_userName :: Lens' UserDetail (Maybe Text)

The friendly name identifying the user.

userDetail_userId :: Lens' UserDetail (Maybe Text)

The stable and unique string identifying the user. For more information about IDs, see IAM identifiers in the IAM User Guide.

userDetail_permissionsBoundary :: Lens' UserDetail (Maybe AttachedPermissionsBoundary)

The ARN of the policy used to set the permissions boundary for the user.

For more information about permissions boundaries, see Permissions boundaries for IAM identities in the IAM User Guide.

userDetail_userPolicyList :: Lens' UserDetail (Maybe [PolicyDetail])

A list of the inline policies embedded in the user.

userDetail_tags :: Lens' UserDetail (Maybe [Tag])

A list of tags that are associated with the user. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

userDetail_attachedManagedPolicies :: Lens' UserDetail (Maybe [AttachedPolicy])

A list of the managed policies attached to the user.