Network.AWS.IAM.Types.VirtualMFADevice

Description

Documentation

data VirtualMFADevice

Contains information about a virtual MFA device.

See: newVirtualMFADevice smart constructor.

Constructors

VirtualMFADevice'

Fields

user :: Maybe User

The IAM user associated with this virtual MFA device.

enableDate :: Maybe ISO8601

The date and time on which the virtual MFA device was enabled.

qRCodePNG :: Maybe (Sensitive Base64)

A QR code PNG image that encodes otpauth://totp/$virtualMFADeviceName@$AccountName?secret=$Base32String where $virtualMFADeviceName is one of the create call arguments. AccountName is the user name if set (otherwise, the account ID otherwise), and Base32String is the seed in base32 format. The Base32String value is base64-encoded.

tags :: Maybe [Tag]

A list of tags that are attached to the virtual MFA device. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

base32StringSeed :: Maybe (Sensitive Base64)

The base32 seed defined as specified in RFC3548. The Base32StringSeed is base64-encoded.

serialNumber :: Text

The serial number associated with VirtualMFADevice.

newVirtualMFADevice

Arguments

:: Text
$sel:serialNumber:VirtualMFADevice'

-> VirtualMFADevice

Create a value of VirtualMFADevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:user:VirtualMFADevice', virtualMFADevice_user - The IAM user associated with this virtual MFA device.

$sel:enableDate:VirtualMFADevice', virtualMFADevice_enableDate - The date and time on which the virtual MFA device was enabled.

$sel:qRCodePNG:VirtualMFADevice', virtualMFADevice_qRCodePNG - A QR code PNG image that encodes otpauth://totp/$virtualMFADeviceName@$AccountName?secret=$Base32String where $virtualMFADeviceName is one of the create call arguments. AccountName is the user name if set (otherwise, the account ID otherwise), and Base32String is the seed in base32 format. The Base32String value is base64-encoded.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:tags:VirtualMFADevice', virtualMFADevice_tags - A list of tags that are attached to the virtual MFA device. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

$sel:base32StringSeed:VirtualMFADevice', virtualMFADevice_base32StringSeed - The base32 seed defined as specified in RFC3548. The Base32StringSeed is base64-encoded.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:serialNumber:VirtualMFADevice', virtualMFADevice_serialNumber - The serial number associated with VirtualMFADevice.

virtualMFADevice_user :: Lens' VirtualMFADevice (Maybe User)

The IAM user associated with this virtual MFA device.

virtualMFADevice_enableDate :: Lens' VirtualMFADevice (Maybe UTCTime)

The date and time on which the virtual MFA device was enabled.

virtualMFADevice_qRCodePNG :: Lens' VirtualMFADevice (Maybe ByteString)

A QR code PNG image that encodes otpauth://totp/$virtualMFADeviceName@$AccountName?secret=$Base32String where $virtualMFADeviceName is one of the create call arguments. AccountName is the user name if set (otherwise, the account ID otherwise), and Base32String is the seed in base32 format. The Base32String value is base64-encoded.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

virtualMFADevice_tags :: Lens' VirtualMFADevice (Maybe [Tag])

A list of tags that are attached to the virtual MFA device. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

virtualMFADevice_base32StringSeed :: Lens' VirtualMFADevice (Maybe ByteString)

The base32 seed defined as specified in RFC3548. The Base32StringSeed is base64-encoded.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

virtualMFADevice_serialNumber :: Lens' VirtualMFADevice Text

The serial number associated with VirtualMFADevice.