Network.AWS.IAM.UntagServerCertificate

Description

Removes the specified tags from the IAM server certificate. For more information about tagging, see Tagging IAM resources in the IAM User Guide.

For certificates in a Region supported by Certificate Manager (ACM), we recommend that you don't use IAM server certificates. Instead, use ACM to provision, manage, and deploy your server certificates. For more information about IAM server certificates, Working with server certificates in the IAM User Guide.

Creating a Request

data UntagServerCertificate

See: newUntagServerCertificate smart constructor.

Constructors

UntagServerCertificate'

Fields

serverCertificateName :: Text

The name of the IAM server certificate from which you want to remove tags.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

tagKeys :: [Text]

A list of key names as a simple array of strings. The tags with matching keys are removed from the specified IAM server certificate.

newUntagServerCertificate

Arguments

:: Text
$sel:serverCertificateName:UntagServerCertificate'

-> UntagServerCertificate

Create a value of UntagServerCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serverCertificateName:UntagServerCertificate', untagServerCertificate_serverCertificateName - The name of the IAM server certificate from which you want to remove tags.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:tagKeys:UntagServerCertificate', untagServerCertificate_tagKeys - A list of key names as a simple array of strings. The tags with matching keys are removed from the specified IAM server certificate.

Request Lenses

untagServerCertificate_serverCertificateName :: Lens' UntagServerCertificate Text

The name of the IAM server certificate from which you want to remove tags.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

untagServerCertificate_tagKeys :: Lens' UntagServerCertificate [Text]

A list of key names as a simple array of strings. The tags with matching keys are removed from the specified IAM server certificate.

Destructuring the Response

data UntagServerCertificateResponse

See: newUntagServerCertificateResponse smart constructor.

Constructors

UntagServerCertificateResponse'

newUntagServerCertificateResponse :: UntagServerCertificateResponse

Create a value of UntagServerCertificateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.