Network.AWS.IAM.UpdateAccessKey

Description

Changes the status of the specified access key from Active to Inactive, or vice versa. This operation can be used to disable a user's key as part of a key rotation workflow.

If the UserName is not specified, the user name is determined implicitly based on the Amazon Web Services access key ID used to sign the request. This operation works for access keys under the Amazon Web Services account. Consequently, you can use this operation to manage Amazon Web Services account root user credentials even if the Amazon Web Services account has no associated users.

For information about rotating keys, see Managing keys and certificates in the IAM User Guide.

Creating a Request

data UpdateAccessKey

See: newUpdateAccessKey smart constructor.

Constructors

UpdateAccessKey'

Fields

userName :: Maybe Text

The name of the user whose key you want to update.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

accessKeyId :: AccessKey

The access key ID of the secret access key you want to update.

This parameter allows (through its regex pattern) a string of characters that can consist of any upper or lowercased letter or digit.

status :: StatusType

The status you want to assign to the secret access key. Active means that the key can be used for programmatic calls to Amazon Web Services, while Inactive means that the key cannot be used.

newUpdateAccessKey

Arguments

:: AccessKey
$sel:accessKeyId:UpdateAccessKey'

-> StatusType
$sel:status:UpdateAccessKey'

-> UpdateAccessKey

Create a value of UpdateAccessKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userName:UpdateAccessKey', updateAccessKey_userName - The name of the user whose key you want to update.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:accessKeyId:UpdateAccessKey', updateAccessKey_accessKeyId - The access key ID of the secret access key you want to update.

This parameter allows (through its regex pattern) a string of characters that can consist of any upper or lowercased letter or digit.

$sel:status:UpdateAccessKey', updateAccessKey_status - The status you want to assign to the secret access key. Active means that the key can be used for programmatic calls to Amazon Web Services, while Inactive means that the key cannot be used.

Request Lenses

updateAccessKey_userName :: Lens' UpdateAccessKey (Maybe Text)

The name of the user whose key you want to update.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

updateAccessKey_accessKeyId :: Lens' UpdateAccessKey AccessKey

The access key ID of the secret access key you want to update.

This parameter allows (through its regex pattern) a string of characters that can consist of any upper or lowercased letter or digit.

updateAccessKey_status :: Lens' UpdateAccessKey StatusType

The status you want to assign to the secret access key. Active means that the key can be used for programmatic calls to Amazon Web Services, while Inactive means that the key cannot be used.

Destructuring the Response

data UpdateAccessKeyResponse

See: newUpdateAccessKeyResponse smart constructor.

Constructors

UpdateAccessKeyResponse'

newUpdateAccessKeyResponse :: UpdateAccessKeyResponse

Create a value of UpdateAccessKeyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.