Network.AWS.IAM.UpdateAccountPasswordPolicy

Description

Updates the password policy settings for the Amazon Web Services account.

This operation does not support partial updates. No parameters are required, but if you do not specify a parameter, that parameter's value reverts to its default value. See the Request Parameters section for each parameter's default value. Also note that some parameters do not allow the default parameter to be explicitly set. Instead, to invoke the default value, do not include that parameter when you invoke the operation.

For more information about using a password policy, see Managing an IAM password policy in the IAM User Guide.

Creating a Request

data UpdateAccountPasswordPolicy

See: newUpdateAccountPasswordPolicy smart constructor.

Constructors

UpdateAccountPasswordPolicy'

Fields

requireLowercaseCharacters :: Maybe Bool

Specifies whether IAM user passwords must contain at least one lowercase character from the ISO basic Latin alphabet (a to z).

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that passwords do not require at least one lowercase character.

maxPasswordAge :: Maybe Natural

The number of days that an IAM user password is valid.

If you do not specify a value for this parameter, then the operation uses the default value of 0. The result is that IAM user passwords never expire.

passwordReusePrevention :: Maybe Natural

Specifies the number of previous passwords that IAM users are prevented from reusing.

If you do not specify a value for this parameter, then the operation uses the default value of 0. The result is that IAM users are not prevented from reusing previous passwords.

minimumPasswordLength :: Maybe Natural

The minimum number of characters allowed in an IAM user password.

If you do not specify a value for this parameter, then the operation uses the default value of 6.

requireUppercaseCharacters :: Maybe Bool

Specifies whether IAM user passwords must contain at least one uppercase character from the ISO basic Latin alphabet (A to Z).

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that passwords do not require at least one uppercase character.

allowUsersToChangePassword :: Maybe Bool

Allows all IAM users in your account to use the Amazon Web Services Management Console to change their own passwords. For more information, see Letting IAM users change their own passwords in the IAM User Guide.

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that IAM users in the account do not automatically have permissions to change their own password.

hardExpiry :: Maybe Bool

Prevents IAM users from setting a new password after their password has expired. The IAM user cannot be accessed until an administrator resets the password.

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that IAM users can change their passwords after they expire and continue to sign in as the user.

requireSymbols :: Maybe Bool

Specifies whether IAM user passwords must contain at least one of the following non-alphanumeric characters:

! @ # $ % ^ & * ( ) _ + - = [ ] { } | '

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that passwords do not require at least one symbol character.

requireNumbers :: Maybe Bool

Specifies whether IAM user passwords must contain at least one numeric character (0 to 9).

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that passwords do not require at least one numeric character.

newUpdateAccountPasswordPolicy :: UpdateAccountPasswordPolicy

Create a value of UpdateAccountPasswordPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requireLowercaseCharacters:UpdateAccountPasswordPolicy', updateAccountPasswordPolicy_requireLowercaseCharacters - Specifies whether IAM user passwords must contain at least one lowercase character from the ISO basic Latin alphabet (a to z).

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that passwords do not require at least one lowercase character.

$sel:maxPasswordAge:UpdateAccountPasswordPolicy', updateAccountPasswordPolicy_maxPasswordAge - The number of days that an IAM user password is valid.

If you do not specify a value for this parameter, then the operation uses the default value of 0. The result is that IAM user passwords never expire.

$sel:passwordReusePrevention:UpdateAccountPasswordPolicy', updateAccountPasswordPolicy_passwordReusePrevention - Specifies the number of previous passwords that IAM users are prevented from reusing.

If you do not specify a value for this parameter, then the operation uses the default value of 0. The result is that IAM users are not prevented from reusing previous passwords.

$sel:minimumPasswordLength:UpdateAccountPasswordPolicy', updateAccountPasswordPolicy_minimumPasswordLength - The minimum number of characters allowed in an IAM user password.

If you do not specify a value for this parameter, then the operation uses the default value of 6.

$sel:requireUppercaseCharacters:UpdateAccountPasswordPolicy', updateAccountPasswordPolicy_requireUppercaseCharacters - Specifies whether IAM user passwords must contain at least one uppercase character from the ISO basic Latin alphabet (A to Z).

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that passwords do not require at least one uppercase character.

$sel:allowUsersToChangePassword:UpdateAccountPasswordPolicy', updateAccountPasswordPolicy_allowUsersToChangePassword - Allows all IAM users in your account to use the Amazon Web Services Management Console to change their own passwords. For more information, see Letting IAM users change their own passwords in the IAM User Guide.

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that IAM users in the account do not automatically have permissions to change their own password.

$sel:hardExpiry:UpdateAccountPasswordPolicy', updateAccountPasswordPolicy_hardExpiry - Prevents IAM users from setting a new password after their password has expired. The IAM user cannot be accessed until an administrator resets the password.

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that IAM users can change their passwords after they expire and continue to sign in as the user.

$sel:requireSymbols:UpdateAccountPasswordPolicy', updateAccountPasswordPolicy_requireSymbols - Specifies whether IAM user passwords must contain at least one of the following non-alphanumeric characters:

! @ # $ % ^ & * ( ) _ + - = [ ] { } | '

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that passwords do not require at least one symbol character.

$sel:requireNumbers:UpdateAccountPasswordPolicy', updateAccountPasswordPolicy_requireNumbers - Specifies whether IAM user passwords must contain at least one numeric character (0 to 9).

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that passwords do not require at least one numeric character.

Request Lenses

updateAccountPasswordPolicy_requireLowercaseCharacters :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Specifies whether IAM user passwords must contain at least one lowercase character from the ISO basic Latin alphabet (a to z).

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that passwords do not require at least one lowercase character.

updateAccountPasswordPolicy_maxPasswordAge :: Lens' UpdateAccountPasswordPolicy (Maybe Natural)

The number of days that an IAM user password is valid.

If you do not specify a value for this parameter, then the operation uses the default value of 0. The result is that IAM user passwords never expire.

updateAccountPasswordPolicy_passwordReusePrevention :: Lens' UpdateAccountPasswordPolicy (Maybe Natural)

Specifies the number of previous passwords that IAM users are prevented from reusing.

If you do not specify a value for this parameter, then the operation uses the default value of 0. The result is that IAM users are not prevented from reusing previous passwords.

updateAccountPasswordPolicy_minimumPasswordLength :: Lens' UpdateAccountPasswordPolicy (Maybe Natural)

The minimum number of characters allowed in an IAM user password.

If you do not specify a value for this parameter, then the operation uses the default value of 6.

updateAccountPasswordPolicy_requireUppercaseCharacters :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Specifies whether IAM user passwords must contain at least one uppercase character from the ISO basic Latin alphabet (A to Z).

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that passwords do not require at least one uppercase character.

updateAccountPasswordPolicy_allowUsersToChangePassword :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Allows all IAM users in your account to use the Amazon Web Services Management Console to change their own passwords. For more information, see Letting IAM users change their own passwords in the IAM User Guide.

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that IAM users in the account do not automatically have permissions to change their own password.

updateAccountPasswordPolicy_hardExpiry :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Prevents IAM users from setting a new password after their password has expired. The IAM user cannot be accessed until an administrator resets the password.

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that IAM users can change their passwords after they expire and continue to sign in as the user.

updateAccountPasswordPolicy_requireSymbols :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Specifies whether IAM user passwords must contain at least one of the following non-alphanumeric characters:

! @ # $ % ^ & * ( ) _ + - = [ ] { } | '

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that passwords do not require at least one symbol character.

updateAccountPasswordPolicy_requireNumbers :: Lens' UpdateAccountPasswordPolicy (Maybe Bool)

Specifies whether IAM user passwords must contain at least one numeric character (0 to 9).

If you do not specify a value for this parameter, then the operation uses the default value of false. The result is that passwords do not require at least one numeric character.

Destructuring the Response

data UpdateAccountPasswordPolicyResponse

See: newUpdateAccountPasswordPolicyResponse smart constructor.

Constructors

UpdateAccountPasswordPolicyResponse'

newUpdateAccountPasswordPolicyResponse :: UpdateAccountPasswordPolicyResponse

Create a value of UpdateAccountPasswordPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.