Network.AWS.IAM.UpdateLoginProfile

Description

Changes the password for the specified IAM user. You can use the CLI, the Amazon Web Services API, or the Users page in the IAM console to change the password for any IAM user. Use ChangePassword to change your own password in the My Security Credentials page in the Amazon Web Services Management Console.

For more information about modifying passwords, see Managing passwords in the IAM User Guide.

Creating a Request

data UpdateLoginProfile

See: newUpdateLoginProfile smart constructor.

Constructors

UpdateLoginProfile'

Fields

passwordResetRequired :: Maybe Bool

Allows this new password to be used only once by requiring the specified IAM user to set a new password on next sign-in.

password :: Maybe (Sensitive Text)

The new password for the specified IAM user.

The regex pattern used to validate this parameter is a string of characters consisting of the following:

Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range

The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)

The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)

However, the format can be further restricted by the account administrator by setting a password policy on the Amazon Web Services account. For more information, see UpdateAccountPasswordPolicy.

userName :: Text

The name of the user whose password you want to update.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

newUpdateLoginProfile

Arguments

:: Text
$sel:userName:UpdateLoginProfile'

-> UpdateLoginProfile

Create a value of UpdateLoginProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:passwordResetRequired:UpdateLoginProfile', updateLoginProfile_passwordResetRequired - Allows this new password to be used only once by requiring the specified IAM user to set a new password on next sign-in.

$sel:password:UpdateLoginProfile', updateLoginProfile_password - The new password for the specified IAM user.

The regex pattern used to validate this parameter is a string of characters consisting of the following:

Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range

The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)

The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)

However, the format can be further restricted by the account administrator by setting a password policy on the Amazon Web Services account. For more information, see UpdateAccountPasswordPolicy.

$sel:userName:UpdateLoginProfile', updateLoginProfile_userName - The name of the user whose password you want to update.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Request Lenses

updateLoginProfile_passwordResetRequired :: Lens' UpdateLoginProfile (Maybe Bool)

Allows this new password to be used only once by requiring the specified IAM user to set a new password on next sign-in.

updateLoginProfile_password :: Lens' UpdateLoginProfile (Maybe Text)

The new password for the specified IAM user.

The regex pattern used to validate this parameter is a string of characters consisting of the following:

Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range

The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)

The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)

However, the format can be further restricted by the account administrator by setting a password policy on the Amazon Web Services account. For more information, see UpdateAccountPasswordPolicy.

updateLoginProfile_userName :: Lens' UpdateLoginProfile Text

The name of the user whose password you want to update.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

Destructuring the Response

data UpdateLoginProfileResponse

See: newUpdateLoginProfileResponse smart constructor.

Constructors

UpdateLoginProfileResponse'

newUpdateLoginProfileResponse :: UpdateLoginProfileResponse

Create a value of UpdateLoginProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.