Network.AWS.IAM.UpdateOpenIDConnectProviderThumbprint

Description

Replaces the existing list of server certificate thumbprints associated with an OpenID Connect (OIDC) provider resource object with a new list of thumbprints.

The list that you pass with this operation completely replaces the existing list of thumbprints. (The lists are not merged.)

Typically, you need to update a thumbprint only when the identity provider certificate changes, which occurs rarely. However, if the provider's certificate does change, any attempt to assume an IAM role that specifies the OIDC provider as a principal fails until the certificate thumbprint is updated.

Amazon Web Services secures communication with some OIDC identity providers (IdPs) through our library of trusted certificate authorities (CAs) instead of using a certificate thumbprint to verify your IdP server certificate. These OIDC IdPs include Google, and those that use an Amazon S3 bucket to host a JSON Web Key Set (JWKS) endpoint. In these cases, your legacy thumbprint remains in your configuration, but is no longer used for validation.

Trust for the OIDC provider is derived from the provider certificate and is validated by the thumbprint. Therefore, it is best to limit access to the UpdateOpenIDConnectProviderThumbprint operation to highly privileged users.

Creating a Request

data UpdateOpenIDConnectProviderThumbprint

See: newUpdateOpenIDConnectProviderThumbprint smart constructor.

Constructors

UpdateOpenIDConnectProviderThumbprint'

Fields

openIDConnectProviderArn :: Text

The Amazon Resource Name (ARN) of the IAM OIDC provider resource object for which you want to update the thumbprint. You can get a list of OIDC provider ARNs by using the ListOpenIDConnectProviders operation.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

thumbprintList :: [Text]

A list of certificate thumbprints that are associated with the specified IAM OpenID Connect provider. For more information, see CreateOpenIDConnectProvider.

newUpdateOpenIDConnectProviderThumbprint

Arguments

:: Text
$sel:openIDConnectProviderArn:UpdateOpenIDConnectProviderThumbprint'

-> UpdateOpenIDConnectProviderThumbprint

Create a value of UpdateOpenIDConnectProviderThumbprint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:openIDConnectProviderArn:UpdateOpenIDConnectProviderThumbprint', updateOpenIDConnectProviderThumbprint_openIDConnectProviderArn - The Amazon Resource Name (ARN) of the IAM OIDC provider resource object for which you want to update the thumbprint. You can get a list of OIDC provider ARNs by using the ListOpenIDConnectProviders operation.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

$sel:thumbprintList:UpdateOpenIDConnectProviderThumbprint', updateOpenIDConnectProviderThumbprint_thumbprintList - A list of certificate thumbprints that are associated with the specified IAM OpenID Connect provider. For more information, see CreateOpenIDConnectProvider.

Request Lenses

updateOpenIDConnectProviderThumbprint_openIDConnectProviderArn :: Lens' UpdateOpenIDConnectProviderThumbprint Text

The Amazon Resource Name (ARN) of the IAM OIDC provider resource object for which you want to update the thumbprint. You can get a list of OIDC provider ARNs by using the ListOpenIDConnectProviders operation.

For more information about ARNs, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

updateOpenIDConnectProviderThumbprint_thumbprintList :: Lens' UpdateOpenIDConnectProviderThumbprint [Text]

A list of certificate thumbprints that are associated with the specified IAM OpenID Connect provider. For more information, see CreateOpenIDConnectProvider.

Destructuring the Response

data UpdateOpenIDConnectProviderThumbprintResponse

See: newUpdateOpenIDConnectProviderThumbprintResponse smart constructor.

Constructors

UpdateOpenIDConnectProviderThumbprintResponse'

newUpdateOpenIDConnectProviderThumbprintResponse :: UpdateOpenIDConnectProviderThumbprintResponse

Create a value of UpdateOpenIDConnectProviderThumbprintResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.