Network.AWS.IAM.UploadSSHPublicKey

Description

Uploads an SSH public key and associates it with the specified IAM user.

The SSH public key uploaded by this operation can be used only for authenticating the associated IAM user to an CodeCommit repository. For more information about using SSH keys to authenticate to an CodeCommit repository, see Set up CodeCommit for SSH connections in the CodeCommit User Guide.

Creating a Request

data UploadSSHPublicKey

See: newUploadSSHPublicKey smart constructor.

Constructors

UploadSSHPublicKey'

Fields

userName :: Text

The name of the IAM user to associate the SSH public key with.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

sSHPublicKeyBody :: Text

The SSH public key. The public key must be encoded in ssh-rsa format or PEM format. The minimum bit-length of the public key is 2048 bits. For example, you can generate a 2048-bit key, and the resulting PEM file is 1679 bytes long.

The regex pattern used to validate this parameter is a string of characters consisting of the following:

Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range

The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)

The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)

newUploadSSHPublicKey

Arguments

:: Text
$sel:userName:UploadSSHPublicKey'

-> Text
$sel:sSHPublicKeyBody:UploadSSHPublicKey'

-> UploadSSHPublicKey

Create a value of UploadSSHPublicKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userName:UploadSSHPublicKey', uploadSSHPublicKey_userName - The name of the IAM user to associate the SSH public key with.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:sSHPublicKeyBody:UploadSSHPublicKey', uploadSSHPublicKey_sSHPublicKeyBody - The SSH public key. The public key must be encoded in ssh-rsa format or PEM format. The minimum bit-length of the public key is 2048 bits. For example, you can generate a 2048-bit key, and the resulting PEM file is 1679 bytes long.

The regex pattern used to validate this parameter is a string of characters consisting of the following:

Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range

The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)

The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)

Request Lenses

uploadSSHPublicKey_userName :: Lens' UploadSSHPublicKey Text

The name of the IAM user to associate the SSH public key with.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

uploadSSHPublicKey_sSHPublicKeyBody :: Lens' UploadSSHPublicKey Text

The SSH public key. The public key must be encoded in ssh-rsa format or PEM format. The minimum bit-length of the public key is 2048 bits. For example, you can generate a 2048-bit key, and the resulting PEM file is 1679 bytes long.

The regex pattern used to validate this parameter is a string of characters consisting of the following:

Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range

The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)

The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)

Destructuring the Response

data UploadSSHPublicKeyResponse

Contains the response to a successful UploadSSHPublicKey request.

See: newUploadSSHPublicKeyResponse smart constructor.

Constructors

UploadSSHPublicKeyResponse'

Fields

sSHPublicKey :: Maybe SSHPublicKey

Contains information about the SSH public key.

httpStatus :: Int

The response's http status code.

newUploadSSHPublicKeyResponse

Arguments

:: Int
$sel:httpStatus:UploadSSHPublicKeyResponse'

-> UploadSSHPublicKeyResponse

Create a value of UploadSSHPublicKeyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sSHPublicKey:UploadSSHPublicKeyResponse', uploadSSHPublicKeyResponse_sSHPublicKey - Contains information about the SSH public key.

$sel:httpStatus:UploadSSHPublicKeyResponse', uploadSSHPublicKeyResponse_httpStatus - The response's http status code.

Response Lenses

uploadSSHPublicKeyResponse_sSHPublicKey :: Lens' UploadSSHPublicKeyResponse (Maybe SSHPublicKey)

Contains information about the SSH public key.

uploadSSHPublicKeyResponse_httpStatus :: Lens' UploadSSHPublicKeyResponse Int

The response's http status code.