Network.AWS.IAM.UploadSigningCertificate

Description

Uploads an X.509 signing certificate and associates it with the specified IAM user. Some Amazon Web Services services require you to use certificates to validate requests that are signed with a corresponding private key. When you upload the certificate, its default status is Active.

For information about when you would use an X.509 signing certificate, see Managing server certificates in IAM in the IAM User Guide.

If the UserName is not specified, the IAM user name is determined implicitly based on the Amazon Web Services access key ID used to sign the request. This operation works for access keys under the Amazon Web Services account. Consequently, you can use this operation to manage Amazon Web Services account root user credentials even if the Amazon Web Services account has no associated users.

Because the body of an X.509 certificate can be large, you should use POST rather than GET when calling UploadSigningCertificate. For information about setting up signatures and authorization through the API, see Signing Amazon Web Services API requests in the Amazon Web Services General Reference. For general information about using the Query API with IAM, see Making query requests in the IAM User Guide.

Creating a Request

data UploadSigningCertificate

See: newUploadSigningCertificate smart constructor.

Constructors

UploadSigningCertificate'

Fields

userName :: Maybe Text

The name of the user the signing certificate is for.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

certificateBody :: Text

The contents of the signing certificate.

The regex pattern used to validate this parameter is a string of characters consisting of the following:

Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range

The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)

The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)

newUploadSigningCertificate

Arguments

:: Text
$sel:certificateBody:UploadSigningCertificate'

-> UploadSigningCertificate

Create a value of UploadSigningCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userName:UploadSigningCertificate', uploadSigningCertificate_userName - The name of the user the signing certificate is for.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

$sel:certificateBody:UploadSigningCertificate', uploadSigningCertificate_certificateBody - The contents of the signing certificate.

The regex pattern used to validate this parameter is a string of characters consisting of the following:

Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range

The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)

The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)

Request Lenses

uploadSigningCertificate_userName :: Lens' UploadSigningCertificate (Maybe Text)

The name of the user the signing certificate is for.

This parameter allows (through its regex pattern) a string of characters consisting of upper and lowercase alphanumeric characters with no spaces. You can also include any of the following characters: _+=,.@-

uploadSigningCertificate_certificateBody :: Lens' UploadSigningCertificate Text

The contents of the signing certificate.

The regex pattern used to validate this parameter is a string of characters consisting of the following:

Any printable ASCII character ranging from the space character (\u0020) through the end of the ASCII character range

The printable characters in the Basic Latin and Latin-1 Supplement character set (through \u00FF)

The special characters tab (\u0009), line feed (\u000A), and carriage return (\u000D)

Destructuring the Response

data UploadSigningCertificateResponse

Contains the response to a successful UploadSigningCertificate request.

See: newUploadSigningCertificateResponse smart constructor.

Constructors

UploadSigningCertificateResponse'

Fields

httpStatus :: Int

The response's http status code.

certificate :: SigningCertificate

Information about the certificate.

newUploadSigningCertificateResponse

Arguments

:: Int
$sel:httpStatus:UploadSigningCertificateResponse'

-> SigningCertificate
$sel:certificate:UploadSigningCertificateResponse'

-> UploadSigningCertificateResponse

Create a value of UploadSigningCertificateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UploadSigningCertificateResponse', uploadSigningCertificateResponse_httpStatus - The response's http status code.

$sel:certificate:UploadSigningCertificateResponse', uploadSigningCertificateResponse_certificate - Information about the certificate.

Response Lenses

uploadSigningCertificateResponse_httpStatus :: Lens' UploadSigningCertificateResponse Int

The response's http status code.

uploadSigningCertificateResponse_certificate :: Lens' UploadSigningCertificateResponse SigningCertificate

Information about the certificate.