Network.AWS.IAM.V2010_05_08.AddRoleToInstanceProfile

Description

Adds the specified role to the specified instance profile. For more information about roles, go to Working with Roles. For more information about instance profiles, go to About Instance Profiles. https:/iam.amazonaws.com ?Action=AddRoleToInstanceProfile &InstanceProfileName=Webserver &RoleName=S3Access &Version=2010-05-08 &AUTHPARAMS 12657608-99f2-11e1-a4c3-27EXAMPLE804.

Documentation

data AddRoleToInstanceProfile

Constructors

AddRoleToInstanceProfile

Fields

_artiprInstanceProfileName :: Text

Name of the instance profile to update.

_artiprRoleName :: Text

Name of the role to add.

artiprRoleName :: Lens' AddRoleToInstanceProfile Text

artiprInstanceProfileName :: Lens' AddRoleToInstanceProfile Text

data AddRoleToInstanceProfileResponse

Constructors

AddRoleToInstanceProfileResponse