Network.AWS.IAM.V2010_05_08.CreateGroup

Description

Creates a new group. For information about the number of groups you can create, see Limitations on IAM Entities in the Using IAM guide. https:/iam.amazonaws.com ?Action=CreateGroup &Path=/ &GroupName=Admins &Version=2010-05-08 &AUTHPARAMS / Admins AGPACKCEVSQ6C2EXAMPLE arn:aws:iam::123456789012:group/Admins 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

createGroup

Arguments

:: Text
_cgrGroupName

-> CreateGroup

Minimum specification for a CreateGroup request.

data CreateGroup

Constructors

CreateGroup

Fields

_cgrGroupName :: Text

Name of the group to create. Do not include the path in this value.

_cgrPath :: Maybe Text

The path to the group. For more information about paths, see Identifiers for IAM Entities in the Using IAM guide. This parameter is optional. If it is not included, it defaults to a slash (/).

cgrPath :: Lens' CreateGroup (Maybe Text)

cgrGroupName :: Lens' CreateGroup Text

data CreateGroupResponse

Constructors

CreateGroupResponse

Fields

_cgsGroup :: Group

Information about the group.

cgsGroup :: Iso' CreateGroupResponse Group