Network.AWS.IAM.V2010_05_08.CreateRole

Description

Creates a new role for your AWS account. For more information about roles, go to Working with Roles. For information about limitations on role names and the number of roles you can create, go to Limitations on IAM Entities in the Using IAM guide. The example policy grants permission to an EC2 instance to assume the role. The policy is URL-encoded according to RFC 3986. For more information about RFC 3986, go to http:/www.faqs.orgrfcsrfc3986.html. https:iam.amazonaws.com ?Action=CreateRole &RoleName=S3Access &Path=application_abccomponent_xyz/ &AssumeRolePolicyDocument={Version:"2012-10-17",Statement:[{Effect:Allow,Principal:{Service:["ec2.amazonaws.com"]},Action:["sts:AssumeRole"]}]} &Version=2010-05-08 &AUTHPARAMS application_abccomponent_xyz/ arn:aws:iam::123456789012:roleapplication_abccomponent_xyz/S3Access S3Access {Version:"2012-10-17",Statement:[{Effect:Allow,Principal:{Service:["ec2.amazonaws.com"]},Action:["sts:AssumeRole"]}]} 2012-05-08T23:34:01.495Z AROADBQP57FF2AEXAMPLE 4a93ceee-9966-11e1-b624-b1aEXAMPLE7c.

Documentation

createRole

Arguments

:: Text
_crrAssumeRolePolicyDocument

-> Text
_crrRoleName

-> CreateRole

Minimum specification for a CreateRole request.

data CreateRole

Constructors

CreateRole

Fields

_crrAssumeRolePolicyDocument :: Text

The policy that grants an entity permission to assume the role.

_crrRoleName :: Text

Name of the role to create.

_crrPath :: Maybe Text

The path to the role. For more information about paths, see Identifiers for IAM Entities in the Using IAM guide. This parameter is optional. If it is not included, it defaults to a slash (/).

crrRoleName :: Lens' CreateRole Text

crrPath :: Lens' CreateRole (Maybe Text)

crrAssumeRolePolicyDocument :: Lens' CreateRole Text

data CreateRoleResponse

Constructors

CreateRoleResponse

Fields

_crsRole :: Role

Information about the role.

crsRole :: Iso' CreateRoleResponse Role