Network.AWS.IAM.V2010_05_08.DeleteAccessKey

Description

Deletes the access key associated with the specified user. If you do not specify a user name, IAM determines the user name implicitly based on the AWS access key ID signing the request. Because this action works for access keys under the AWS account, you can use this API to manage root credentials even if the AWS account has no associated users. https:/iam.amazonaws.com ?Action=DeleteAccessKey &UserName=Bob &AccessKeyId=AKIAIOSFODNN7EXAMPLE &Version=2010-05-08 &AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

deleteAccessKey

Arguments

:: Text
_dakrAccessKeyId

-> DeleteAccessKey

Minimum specification for a DeleteAccessKey request.

data DeleteAccessKey

Constructors

DeleteAccessKey

Fields

_dakrAccessKeyId :: Text

The access key ID for the access key ID and secret access key you want to delete.

_dakrUserName :: Maybe Text

Name of the user whose key you want to delete.

dakrUserName :: Lens' DeleteAccessKey (Maybe Text)

dakrAccessKeyId :: Lens' DeleteAccessKey Text

data DeleteAccessKeyResponse

Constructors

DeleteAccessKeyResponse