Network.AWS.IAM.V2010_05_08.DeleteInstanceProfile

Description

Deletes the specified instance profile. The instance profile must not have an associated role. Make sure you do not have any Amazon EC2 instances running with the instance profile you are about to delete. Deleting a role or instance profile that is associated with a running instance will break any applications running on the instance. For more information about instance profiles, go to About Instance Profiles. https:/iam.amazonaws.com ?Action=DeleteInstanceProfile &InstanceProfileName=Webserver &Version=2010-05-08 &AUTHPARAMS 90c18667-99f3-11e1-a4c3-27EXAMPLE804.

Documentation

data DeleteInstanceProfile

Constructors

DeleteInstanceProfile

Fields

_diprInstanceProfileName :: Text

Name of the instance profile to delete.

diprInstanceProfileName :: Iso' DeleteInstanceProfile Text

data DeleteInstanceProfileResponse

Constructors

DeleteInstanceProfileResponse