Network.AWS.IAM.V2010_05_08.DeleteServerCertificate

Description

Deletes the specified server certificate. If you are using a server certificate with Elastic Load Balancing, deleting the certificate could have implications for your application. If Elastic Load Balancing doesn't detect the deletion of bound certificates, it may continue to use the certificates. This could cause Elastic Load Balancing to stop accepting traffic. We recommend that you remove the reference to the certificate from Elastic Load Balancing before using this command to delete the certificate. For more information, go to DeleteLoadBalancerListeners in the Elastic Load Balancing API Reference. https:/iam.amazonaws.com ?Action=DeleteServerCertificate &ServerCertificateName=ProdServerCert &Version=2010-05-08 &AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

data DeleteServerCertificate

Constructors

DeleteServerCertificate

Fields

_dscsServerCertificateName :: Text

The name of the server certificate you want to delete.

dscsServerCertificateName :: Iso' DeleteServerCertificate Text

data DeleteServerCertificateResponse

Constructors

DeleteServerCertificateResponse