Network.AWS.IAM.V2010_05_08.GetCredentialReport

Description

Retrieves a credential report for the AWS account. For more information about the credential report, see Getting Credential Reports in the Using IAM guide.

Documentation

data GetCredentialReport

Constructors

GetCredentialReport

data GetCredentialReportResponse

Constructors

GetCredentialReportResponse

Fields

_gcrrGeneratedTime :: Maybe ISO8601

The time and date when the credential report was created, in ISO 8601 date-time format.

_gcrrContent :: Maybe ByteString

Contains the credential report. The report is Base64-encoded.

_gcrrReportFormat :: Maybe ReportFormatType

The format (MIME type) of the credential report.

gcrrReportFormat :: Lens' GetCredentialReportResponse (Maybe ReportFormatType)

gcrrGeneratedTime :: Lens' GetCredentialReportResponse (Maybe ISO8601)

gcrrContent :: Lens' GetCredentialReportResponse (Maybe ByteString)