Network.AWS.IAM.V2010_05_08.GetGroup

Description

Returns a list of users that are in the specified group. You can paginate the results using the MaxItems and Marker parameters. https:/iam.amazonaws.com ?Action=GetGroup &GroupName=Admins &Version=2010-05-08 &AUTHPARAMS / Admins AGPACKCEVSQ6C2EXAMPLE arn:aws:iam::123456789012:groupAdmins division_abcsubdivision_xyz Bob AIDACKCEVSQ6C2EXAMPLE arn:aws:iam::123456789012:userdivision_abcsubdivision_xyz/Bob division_abcsubdivision_xyz/ Susan AIDACKCEVSQ6C2EXAMPLE arn:aws:iam::123456789012:userdivision_abcsubdivision_xyz/Susan false 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

getGroup

Arguments

:: Text
_ggrGroupName

-> GetGroup

Minimum specification for a GetGroup request.

data GetGroup

Constructors

GetGroup

Fields

_ggrGroupName :: Text

Name of the group.

_ggrMarker :: Maybe Text

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

_ggrMaxItems :: Maybe Integer

Use this only when paginating results to indicate the maximum number of user names you want in the response. If there are additional user names beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

ggrMaxItems :: Lens' GetGroup (Maybe Integer)

ggrMarker :: Lens' GetGroup (Maybe Text)

ggrGroupName :: Lens' GetGroup Text

data GetGroupResponse

Constructors

GetGroupResponse

Fields

_ggsIsTruncated :: Bool

A flag that indicates whether there are more user names to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more user names in the list.

_ggsGroup :: Group

Information about the group.

_ggsUsers :: [User]

A list of users in the group.

_ggsMarker :: Maybe Text

If IsTruncated is true, then this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

ggsUsers :: Lens' GetGroupResponse [User]

ggsMarker :: Lens' GetGroupResponse (Maybe Text)

ggsIsTruncated :: Lens' GetGroupResponse Bool

ggsGroup :: Lens' GetGroupResponse Group