Network.AWS.IAM.V2010_05_08.GetRole

Description

Retrieves information about the specified role, including the role's path, GUID, ARN, and the policy granting permission to assume the role. For more information about ARNs, go to ARNs. For more information about roles, go to Working with Roles. The returned policy is URL-encoded according to RFC 3986. For more information about RFC 3986, go to http:/www.faqs.orgrfcsrfc3986.html. https:iam.amazonaws.com ?Action=GetRole &RoleName=S3Access &Version=2010-05-08 &AUTHPARAMS application_abccomponent_xyz/ arn:aws:iam::123456789012:roleapplication_abccomponent_xyz/S3Access S3Access {Version:"2012-10-17",Statement:[{Effect:Allow,Principal:{Service:["ec2.amazonaws.com"]},Action:["sts:AssumeRole"]}]} 2012-05-08T23:34:01Z AROADBQP57FF2AEXAMPLE df37e965-9967-11e1-a4c3-270EXAMPLE04.

Documentation

data GetRole

Constructors

GetRole

Fields

_grrRoleName :: Text

Name of the role to get information about.

grrRoleName :: Iso' GetRole Text

data GetRoleResponse

Constructors

GetRoleResponse

Fields

_grsRole :: Role

Information about the role.

grsRole :: Iso' GetRoleResponse Role