Network.AWS.IAM.V2010_05_08.GetRolePolicy

Description

Retrieves the specified policy document for the specified role. For more information about roles, go to Working with Roles. The returned policy is URL-encoded according to RFC 3986. For more information about RFC 3986, go to http:/www.faqs.orgrfcsrfc3986.html. https:iam.amazonaws.com ?Action=GetRolePolicy &PolicyName=S3AccessPolicy &RoleName=S3Access &Version=2010-05-08 &AUTHPARAMS S3AccessPolicy S3Access {Version:"2012-10-17",Statement:[{Effect:Allow,Action:["s3:*"],Resource:["*"]}]} 7e7cd8bc-99ef-11e1-a4c3-27EXAMPLE804.

Documentation

data GetRolePolicy

Constructors

GetRolePolicy

Fields

_grprPolicyName :: Text

Name of the policy document to get.

_grprRoleName :: Text

Name of the role associated with the policy.

grprRoleName :: Lens' GetRolePolicy Text

grprPolicyName :: Lens' GetRolePolicy Text

data GetRolePolicyResponse

Constructors

GetRolePolicyResponse

Fields

_grpsPolicyDocument :: Text

The policy document.

_grpsPolicyName :: Text

The name of the policy.

_grpsRoleName :: Text

The role the policy is associated with.

grpsRoleName :: Lens' GetRolePolicyResponse Text

grpsPolicyName :: Lens' GetRolePolicyResponse Text

grpsPolicyDocument :: Lens' GetRolePolicyResponse Text