Network.AWS.IAM.V2010_05_08.GetSAMLProvider

Description

Returns the SAML provider metadocument that was uploaded when the provider was created or updated. This operation requires Signature Version 4. https:/iam.amazonaws.com ?Action=GetSAMLProvider &Name=arn:aws:iam::123456789012:saml-metadata/MyUniversity &Version=2010-05-08 &AUTHPARAMS 2012-05-09T16:27:11Z 2015-12-31T211:59:59Z Pd9fexDssTkRgGNqs...DxptfEs== 29f47818-99f5-11e1-a4c3-27EXAMPLE804.

Documentation

data GetSAMLProvider

Constructors

GetSAMLProvider

Fields

_gsamlprSAMLProviderArn :: Text

The Amazon Resource Name (ARN) of the SAML provider to get information about.

gsamlprSAMLProviderArn :: Iso' GetSAMLProvider Text

data GetSAMLProviderResponse

Constructors

GetSAMLProviderResponse

Fields

_gsamlpsCreateDate :: Maybe ISO8601

The date and time when the SAML provider was created.

_gsamlpsValidUntil :: Maybe ISO8601

The expiration date and time for the SAML provider.

_gsamlpsSAMLMetadataDocument :: Maybe Text

The XML metadata document that includes information about an identity provider.

gsamlpsValidUntil :: Lens' GetSAMLProviderResponse (Maybe ISO8601)

gsamlpsSAMLMetadataDocument :: Lens' GetSAMLProviderResponse (Maybe Text)

gsamlpsCreateDate :: Lens' GetSAMLProviderResponse (Maybe ISO8601)