Network.AWS.IAM.V2010_05_08.GetUser

Description

Retrieves information about the specified user, including the user's path, unique ID, and ARN. If you do not specify a user name, IAM determines the user name implicitly based on the AWS access key ID signing the request. https:/iam.amazonaws.com ?Action=GetUser &UserName=Bob &Version=2010-05-08 &AUTHPARAMS division_abcsubdivision_xyz/ Bob AIDACKCEVSQ6C2EXAMPLE arn:aws:iam::123456789012:userdivision_abcsubdivision_xyz/Bob 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

getUser :: GetUser

Minimum specification for a GetUser request.

data GetUser

Constructors

GetUser

Fields

_gurUserName :: Maybe Text

Name of the user to get information about. This parameter is optional. If it is not included, it defaults to the user making the request.

gurUserName :: Iso' GetUser (Maybe Text)

data GetUserResponse

Constructors

GetUserResponse

Fields

_gusUser :: User

Information about the user.

gusUser :: Iso' GetUserResponse User