Network.AWS.IAM.V2010_05_08.ListAccountAliases

Description

Lists the account aliases associated with the account. For information about using an AWS account alias, see Using an Alias for Your AWS Account ID in the Using IAM guide. You can paginate the results using the MaxItems and Marker parameters. https:/iam.amazonaws.com ?Action=ListAccountAliases &Version=2010-05-08 &AUTHPARAMS false foocorporation c5a076e9-f1b0-11df-8fbe-45274EXAMPLE.

Documentation

listAccountAliases :: ListAccountAliases

Minimum specification for a ListAccountAliases request.

data ListAccountAliases

Constructors

ListAccountAliases

Fields

_laarMarker :: Maybe Text

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

_laarMaxItems :: Maybe Integer

Use this only when paginating results to indicate the maximum number of account aliases you want in the response. If there are additional account aliases beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

laarMaxItems :: Lens' ListAccountAliases (Maybe Integer)

laarMarker :: Lens' ListAccountAliases (Maybe Text)

data ListAccountAliasesResponse

Constructors

ListAccountAliasesResponse

Fields

_laasAccountAliases :: [Text]

A list of aliases associated with the account.

_laasIsTruncated :: Bool

A flag that indicates whether there are more account aliases to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more account aliases in the list.

_laasMarker :: Maybe Text

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

laasMarker :: Lens' ListAccountAliasesResponse (Maybe Text)

laasIsTruncated :: Lens' ListAccountAliasesResponse Bool

laasAccountAliases :: Lens' ListAccountAliasesResponse [Text]