Network.AWS.IAM.V2010_05_08.ListGroupsForUser

Description

Lists the groups the specified user belongs to. You can paginate the results using the MaxItems and Marker parameters. https:/iam.amazonaws.com ?Action=ListGroupsForUser &UserName=Bob &AUTHPARAMS / Admins AGPACKCEVSQ6C2EXAMPLE arn:aws:iam::123456789012:group/Admins false 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

listGroupsForUser

Arguments

:: Text
_lgfurUserName

-> ListGroupsForUser

Minimum specification for a ListGroupsForUser request.

data ListGroupsForUser

Constructors

ListGroupsForUser

Fields

_lgfurUserName :: Text

The name of the user to list groups for.

_lgfurMarker :: Maybe Text

Use this only when paginating results, and only in a subsequent request after you've received a response where the results are truncated. Set it to the value of the Marker element in the response you just received.

_lgfurMaxItems :: Maybe Integer

Use this only when paginating results to indicate the maximum number of groups you want in the response. If there are additional groups beyond the maximum you specify, the IsTruncated response element is true. This parameter is optional. If you do not include it, it defaults to 100.

lgfurUserName :: Lens' ListGroupsForUser Text

lgfurMaxItems :: Lens' ListGroupsForUser (Maybe Integer)

lgfurMarker :: Lens' ListGroupsForUser (Maybe Text)

data ListGroupsForUserResponse

Constructors

ListGroupsForUserResponse

Fields

_lgfusIsTruncated :: Bool

A flag that indicates whether there are more groups to list. If your results were truncated, you can make a subsequent pagination request using the Marker request parameter to retrieve more groups in the list.

_lgfusGroups :: [Group]

A list of groups.

_lgfusMarker :: Maybe Text

If IsTruncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent pagination request.

lgfusMarker :: Lens' ListGroupsForUserResponse (Maybe Text)

lgfusIsTruncated :: Lens' ListGroupsForUserResponse Bool

lgfusGroups :: Lens' ListGroupsForUserResponse [Group]