Network.AWS.IAM.V2010_05_08.RemoveRoleFromInstanceProfile

Description

Removes the specified role from the specified instance profile. Make sure you do not have any Amazon EC2 instances running with the role you are about to remove from the instance profile. Removing a role from an instance profile that is associated with a running instance will break any applications running on the instance. For more information about roles, go to Working with Roles. For more information about instance profiles, go to About Instance Profiles. https:/iam.amazonaws.com ?Action=RemoveRoleFromInstanceProfile &InstanceProfileName=Webserver &RoleName=S3Access &Version=2010-05-08 &AUTHPARAMS 29f47818-99f5-11e1-a4c3-27EXAMPLE804.

Documentation

data RemoveRoleFromInstanceProfile

Constructors

RemoveRoleFromInstanceProfile

Fields

_rrfiprInstanceProfileName :: Text

Name of the instance profile to update.

_rrfiprRoleName :: Text

Name of the role to remove.

rrfiprRoleName :: Lens' RemoveRoleFromInstanceProfile Text

rrfiprInstanceProfileName :: Lens' RemoveRoleFromInstanceProfile Text

data RemoveRoleFromInstanceProfileResponse

Constructors

RemoveRoleFromInstanceProfileResponse