Network.AWS.IAM.V2010_05_08.UpdateGroup

Description

Updates the name and/or the path of the specified group. You should understand the implications of changing a group's path or name. For more information, see Renaming Users and Groups in the Using IAM guide. To change a group name the requester must have appropriate permissions on both the source object and the target object. For example, to change Managers to MGRs, the entity making the request must have permission on Managers and MGRs, or must have permission on all (*). For more information about permissions, see Permissions and Policies. https:/iam.amazonaws.com ?Action=UpdateGroup &GroupName=Test &NewGroupName=Test_1 &Version=2010-05-08 &AUTHPARAMS division_abcsubdivision_xyzproduct_1234engineering/ Test_1 AGP2MAB8DPLSRHEXAMPLE arn:aws:iam::123456789012:groupdivision_abcsubdivision_xyz/ product_1234engineeringTest_1 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

updateGroup

Arguments

:: Text
_ugrGroupName

-> UpdateGroup

Minimum specification for a UpdateGroup request.

data UpdateGroup

Constructors

UpdateGroup

Fields

_ugrGroupName :: Text

Name of the group to update. If you're changing the name of the group, this is the original name.

_ugrNewGroupName :: Maybe Text

New name for the group. Only include this if changing the group's name.

_ugrNewPath :: Maybe Text

New path for the group. Only include this if changing the group's path.

ugrNewPath :: Lens' UpdateGroup (Maybe Text)

ugrNewGroupName :: Lens' UpdateGroup (Maybe Text)

ugrGroupName :: Lens' UpdateGroup Text

data UpdateGroupResponse

Constructors

UpdateGroupResponse