Network.AWS.IAM.V2010_05_08.UpdateLoginProfile

Description

Changes the password for the specified user. https:/iam.amazonaws.com ?Action=UpdateLoginProfile &UserName=Bob &Password=NewPassword &AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE.

Documentation

updateLoginProfile

Arguments

:: Text
_ulprUserName

-> UpdateLoginProfile

Minimum specification for a UpdateLoginProfile request.

data UpdateLoginProfile

Constructors

UpdateLoginProfile

Fields

_ulprUserName :: Text

Name of the user whose password you want to update.

_ulprPasswordResetRequired :: Maybe Bool

Require the specified user to set a new password on next sign-in.

_ulprPassword :: Maybe Text

The new password for the specified user.

ulprUserName :: Lens' UpdateLoginProfile Text

ulprPasswordResetRequired :: Lens' UpdateLoginProfile (Maybe Bool)

ulprPassword :: Lens' UpdateLoginProfile (Maybe Text)

data UpdateLoginProfileResponse

Constructors

UpdateLoginProfileResponse