Network.AWS.IAM.V2010_05_08.UpdateSAMLProvider

Description

Updates the metadata document for an existing SAML provider. This operation requires Signature Version 4. https:/iam.amazonaws.com ?Action=UpdateSAMLProvider &Name=arn:aws:iam::123456789012:saml-metadata/MyUniversity &SAMLProviderDocument=VGhpcyBpcyB3aGVyZSB5b3UgcHV0IHRoZSBTQU1MIHByb3ZpZGVyIG1ldGFkYXRhIGRvY3VtZW50 LCBCYXNlNjQtZW5jb2RlZCBpbnRvIGEgYmlnIHN0cmluZy4= &Version=2010-05-08 &AUTHPARAMS arn:aws:iam::123456789012:saml-metadata/MyUniversity 29f47818-99f5-11e1-a4c3-27EXAMPLE804.

Documentation

data UpdateSAMLProvider

Constructors

UpdateSAMLProvider

Fields

_usamlprSAMLProviderArn :: Text

The Amazon Resource Name (ARN) of the SAML provider to update.

_usamlprSAMLMetadataDocument :: Text

An XML document generated by an identity provider (IdP) that supports SAML 2.0. The document includes the issuer's name, expiration information, and keys that can be used to validate the SAML authentication response (assertions) that are received from the IdP. You must generate the metadata document using the identity management software that is used as your organization's IdP.

usamlprSAMLProviderArn :: Lens' UpdateSAMLProvider Text

usamlprSAMLMetadataDocument :: Lens' UpdateSAMLProvider Text

data UpdateSAMLProviderResponse

Constructors

UpdateSAMLProviderResponse

Fields

_usamlpsSAMLProviderArn :: Maybe Text

The Amazon Resource Name (ARN) of the SAML provider that was updated.

usamlpsSAMLProviderArn :: Iso' UpdateSAMLProviderResponse (Maybe Text)