Network.AWS.IVS.CreateStreamKey

Description

Creates a stream key, used to initiate a stream, for the specified channel ARN.

Note that CreateChannel creates a stream key. If you subsequently use CreateStreamKey on the same channel, it will fail because a stream key already exists and there is a limit of 1 stream key per channel. To reset the stream key on a channel, use DeleteStreamKey and then CreateStreamKey.

Creating a Request

data CreateStreamKey

See: newCreateStreamKey smart constructor.

Constructors

CreateStreamKey'

Fields

tags :: Maybe (HashMap Text Text)

Array of 1-50 maps, each of the form string:string (key:value).

channelArn :: Text

ARN of the channel for which to create the stream key.

newCreateStreamKey

Arguments

:: Text
$sel:channelArn:CreateStreamKey'

-> CreateStreamKey

Create a value of CreateStreamKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateStreamKey', createStreamKey_tags - Array of 1-50 maps, each of the form string:string (key:value).

$sel:channelArn:CreateStreamKey', createStreamKey_channelArn - ARN of the channel for which to create the stream key.

Request Lenses

createStreamKey_tags :: Lens' CreateStreamKey (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

createStreamKey_channelArn :: Lens' CreateStreamKey Text

ARN of the channel for which to create the stream key.

Destructuring the Response

data CreateStreamKeyResponse

See: newCreateStreamKeyResponse smart constructor.

Constructors

CreateStreamKeyResponse'

Fields

streamKey :: Maybe StreamKey

Stream key used to authenticate an RTMPS stream for ingestion.

httpStatus :: Int

The response's http status code.

newCreateStreamKeyResponse

Arguments

:: Int
$sel:httpStatus:CreateStreamKeyResponse'

-> CreateStreamKeyResponse

Create a value of CreateStreamKeyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamKey:CreateStreamKeyResponse', createStreamKeyResponse_streamKey - Stream key used to authenticate an RTMPS stream for ingestion.

$sel:httpStatus:CreateStreamKeyResponse', createStreamKeyResponse_httpStatus - The response's http status code.

Response Lenses

createStreamKeyResponse_streamKey :: Lens' CreateStreamKeyResponse (Maybe StreamKey)

Stream key used to authenticate an RTMPS stream for ingestion.

createStreamKeyResponse_httpStatus :: Lens' CreateStreamKeyResponse Int

The response's http status code.