Network.AWS.IVS.DeleteRecordingConfiguration

Description

Deletes the recording configuration for the specified ARN.

If you try to delete a recording configuration that is associated with a channel, you will get an error (409 ConflictException). To avoid this, for all channels that reference the recording configuration, first use UpdateChannel to set the recordingConfigurationArn field to an empty string, then use DeleteRecordingConfiguration.

Creating a Request

data DeleteRecordingConfiguration

See: newDeleteRecordingConfiguration smart constructor.

Constructors

DeleteRecordingConfiguration'

Fields

arn :: Text

ARN of the recording configuration to be deleted.

newDeleteRecordingConfiguration

Arguments

:: Text
$sel:arn:DeleteRecordingConfiguration'

-> DeleteRecordingConfiguration

Create a value of DeleteRecordingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:DeleteRecordingConfiguration', deleteRecordingConfiguration_arn - ARN of the recording configuration to be deleted.

Request Lenses

deleteRecordingConfiguration_arn :: Lens' DeleteRecordingConfiguration Text

ARN of the recording configuration to be deleted.

Destructuring the Response

data DeleteRecordingConfigurationResponse

See: newDeleteRecordingConfigurationResponse smart constructor.

Constructors

DeleteRecordingConfigurationResponse'

newDeleteRecordingConfigurationResponse :: DeleteRecordingConfigurationResponse

Create a value of DeleteRecordingConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.