Network.AWS.IVS.ListRecordingConfigurations

Description

Gets summary information about all recording configurations in your account, in the Amazon Web Services region where the API request is processed.

This operation returns paginated results.

Creating a Request

data ListRecordingConfigurations

See: newListRecordingConfigurations smart constructor.

Constructors

ListRecordingConfigurations'

Fields

nextToken :: Maybe Text

The first recording configuration to retrieve. This is used for pagination; see the nextToken response field.

maxResults :: Maybe Natural

Maximum number of recording configurations to return. Default: 50.

newListRecordingConfigurations :: ListRecordingConfigurations

Create a value of ListRecordingConfigurations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRecordingConfigurations', listRecordingConfigurations_nextToken - The first recording configuration to retrieve. This is used for pagination; see the nextToken response field.

$sel:maxResults:ListRecordingConfigurations', listRecordingConfigurations_maxResults - Maximum number of recording configurations to return. Default: 50.

Request Lenses

listRecordingConfigurations_nextToken :: Lens' ListRecordingConfigurations (Maybe Text)

The first recording configuration to retrieve. This is used for pagination; see the nextToken response field.

listRecordingConfigurations_maxResults :: Lens' ListRecordingConfigurations (Maybe Natural)

Maximum number of recording configurations to return. Default: 50.

Destructuring the Response

data ListRecordingConfigurationsResponse

See: newListRecordingConfigurationsResponse smart constructor.

Constructors

ListRecordingConfigurationsResponse'

Fields

nextToken :: Maybe Text

If there are more recording configurations than maxResults, use nextToken in the request to get the next set.

httpStatus :: Int

The response's http status code.

recordingConfigurations :: [RecordingConfigurationSummary]

List of the matching recording configurations.

newListRecordingConfigurationsResponse

Arguments

:: Int
$sel:httpStatus:ListRecordingConfigurationsResponse'

-> ListRecordingConfigurationsResponse

Create a value of ListRecordingConfigurationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRecordingConfigurations', listRecordingConfigurationsResponse_nextToken - If there are more recording configurations than maxResults, use nextToken in the request to get the next set.

$sel:httpStatus:ListRecordingConfigurationsResponse', listRecordingConfigurationsResponse_httpStatus - The response's http status code.

$sel:recordingConfigurations:ListRecordingConfigurationsResponse', listRecordingConfigurationsResponse_recordingConfigurations - List of the matching recording configurations.

Response Lenses

listRecordingConfigurationsResponse_nextToken :: Lens' ListRecordingConfigurationsResponse (Maybe Text)

If there are more recording configurations than maxResults, use nextToken in the request to get the next set.

listRecordingConfigurationsResponse_httpStatus :: Lens' ListRecordingConfigurationsResponse Int

The response's http status code.

listRecordingConfigurationsResponse_recordingConfigurations :: Lens' ListRecordingConfigurationsResponse [RecordingConfigurationSummary]

List of the matching recording configurations.