Network.AWS.IVS.Types.Channel

Description

Documentation

data Channel

Object specifying a channel.

See: newChannel smart constructor.

Constructors

Channel'

Fields

playbackUrl :: Maybe Text

Channel playback URL.

authorized :: Maybe Bool

Whether the channel is private (enabled for playback authorization). Default: false.

arn :: Maybe Text

Channel ARN.

latencyMode :: Maybe ChannelLatencyMode

Channel latency mode. Use NORMAL to broadcast and deliver live video up to Full HD. Use LOW for near-real-time interaction with viewers. Default: LOW. (Note: In the Amazon IVS console, LOW and NORMAL correspond to Ultra-low and Standard, respectively.)

name :: Maybe Text

Channel name.

recordingConfigurationArn :: Maybe Text

Recording-configuration ARN. A value other than an empty string indicates that recording is enabled. Default: "" (empty string, recording is disabled).

type' :: Maybe ChannelType

Channel type, which determines the allowable resolution and bitrate. /If you exceed the allowable resolution or bitrate, the stream probably will disconnect immediately./ Default: STANDARD. Valid values:

STANDARD: Multiple qualities are generated from the original input, to automatically give viewers the best experience for their devices and network conditions. Resolution can be up to 1080p and bitrate can be up to 8.5 Mbps. Audio is transcoded only for renditions 360p and below; above that, audio is passed through.

BASIC: Amazon IVS delivers the original input to viewers. The viewer’s video-quality choice is limited to the original input. Resolution can be up to 480p and bitrate can be up to 1.5 Mbps.

tags :: Maybe (HashMap Text Text)

Array of 1-50 maps, each of the form string:string (key:value).

ingestEndpoint :: Maybe Text

Channel ingest endpoint, part of the definition of an ingest server, used when you set up streaming software.

newChannel :: Channel

Create a value of Channel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:playbackUrl:Channel', channel_playbackUrl - Channel playback URL.

$sel:authorized:Channel', channel_authorized - Whether the channel is private (enabled for playback authorization). Default: false.

$sel:arn:Channel', channel_arn - Channel ARN.

$sel:latencyMode:Channel', channel_latencyMode - Channel latency mode. Use NORMAL to broadcast and deliver live video up to Full HD. Use LOW for near-real-time interaction with viewers. Default: LOW. (Note: In the Amazon IVS console, LOW and NORMAL correspond to Ultra-low and Standard, respectively.)

$sel:name:Channel', channel_name - Channel name.

$sel:recordingConfigurationArn:Channel', channel_recordingConfigurationArn - Recording-configuration ARN. A value other than an empty string indicates that recording is enabled. Default: "" (empty string, recording is disabled).

$sel:type':Channel', channel_type - Channel type, which determines the allowable resolution and bitrate. /If you exceed the allowable resolution or bitrate, the stream probably will disconnect immediately./ Default: STANDARD. Valid values:

STANDARD: Multiple qualities are generated from the original input, to automatically give viewers the best experience for their devices and network conditions. Resolution can be up to 1080p and bitrate can be up to 8.5 Mbps. Audio is transcoded only for renditions 360p and below; above that, audio is passed through.

BASIC: Amazon IVS delivers the original input to viewers. The viewer’s video-quality choice is limited to the original input. Resolution can be up to 480p and bitrate can be up to 1.5 Mbps.

$sel:tags:Channel', channel_tags - Array of 1-50 maps, each of the form string:string (key:value).

$sel:ingestEndpoint:Channel', channel_ingestEndpoint - Channel ingest endpoint, part of the definition of an ingest server, used when you set up streaming software.

channel_playbackUrl :: Lens' Channel (Maybe Text)

Channel playback URL.

channel_authorized :: Lens' Channel (Maybe Bool)

Whether the channel is private (enabled for playback authorization). Default: false.

channel_arn :: Lens' Channel (Maybe Text)

Channel ARN.

channel_latencyMode :: Lens' Channel (Maybe ChannelLatencyMode)

Channel latency mode. Use NORMAL to broadcast and deliver live video up to Full HD. Use LOW for near-real-time interaction with viewers. Default: LOW. (Note: In the Amazon IVS console, LOW and NORMAL correspond to Ultra-low and Standard, respectively.)

channel_name :: Lens' Channel (Maybe Text)

Channel name.

channel_recordingConfigurationArn :: Lens' Channel (Maybe Text)

Recording-configuration ARN. A value other than an empty string indicates that recording is enabled. Default: "" (empty string, recording is disabled).

channel_type :: Lens' Channel (Maybe ChannelType)

Channel type, which determines the allowable resolution and bitrate. /If you exceed the allowable resolution or bitrate, the stream probably will disconnect immediately./ Default: STANDARD. Valid values:

STANDARD: Multiple qualities are generated from the original input, to automatically give viewers the best experience for their devices and network conditions. Resolution can be up to 1080p and bitrate can be up to 8.5 Mbps. Audio is transcoded only for renditions 360p and below; above that, audio is passed through.

BASIC: Amazon IVS delivers the original input to viewers. The viewer’s video-quality choice is limited to the original input. Resolution can be up to 480p and bitrate can be up to 1.5 Mbps.

channel_tags :: Lens' Channel (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

channel_ingestEndpoint :: Lens' Channel (Maybe Text)

Channel ingest endpoint, part of the definition of an ingest server, used when you set up streaming software.