Network.AWS.IVS.Types.PlaybackKeyPair

Description

Documentation

data PlaybackKeyPair

A key pair used to sign and validate a playback authorization token.

See: newPlaybackKeyPair smart constructor.

Constructors

PlaybackKeyPair'

Fields

arn :: Maybe Text

Key-pair ARN.

fingerprint :: Maybe Text

Key-pair identifier.

name :: Maybe Text

Playback-key-pair name. The value does not need to be unique.

tags :: Maybe (HashMap Text Text)

Array of 1-50 maps, each of the form string:string (key:value).

newPlaybackKeyPair :: PlaybackKeyPair

Create a value of PlaybackKeyPair with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:PlaybackKeyPair', playbackKeyPair_arn - Key-pair ARN.

$sel:fingerprint:PlaybackKeyPair', playbackKeyPair_fingerprint - Key-pair identifier.

$sel:name:PlaybackKeyPair', playbackKeyPair_name - Playback-key-pair name. The value does not need to be unique.

$sel:tags:PlaybackKeyPair', playbackKeyPair_tags - Array of 1-50 maps, each of the form string:string (key:value).

playbackKeyPair_arn :: Lens' PlaybackKeyPair (Maybe Text)

Key-pair ARN.

playbackKeyPair_fingerprint :: Lens' PlaybackKeyPair (Maybe Text)

Key-pair identifier.

playbackKeyPair_name :: Lens' PlaybackKeyPair (Maybe Text)

Playback-key-pair name. The value does not need to be unique.

playbackKeyPair_tags :: Lens' PlaybackKeyPair (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).