Network.AWS.IVS.Types.RecordingConfiguration

Description

Documentation

data RecordingConfiguration

An object representing a configuration to record a channel stream.

See: newRecordingConfiguration smart constructor.

Constructors

RecordingConfiguration'

Fields

name :: Maybe Text

Recording-configuration name. The value does not need to be unique.

tags :: Maybe (HashMap Text Text)

Array of 1-50 maps, each of the form string:string (key:value).

arn :: Text

Recording-configuration ARN.

destinationConfiguration :: DestinationConfiguration

A complex type that contains information about where recorded video will be stored.

state :: RecordingConfigurationState

Indicates the current state of the recording configuration. When the state is ACTIVE, the configuration is ready for recording a channel stream.

newRecordingConfiguration

Arguments

:: Text
$sel:arn:RecordingConfiguration'

-> DestinationConfiguration
$sel:destinationConfiguration:RecordingConfiguration'

-> RecordingConfigurationState
$sel:state:RecordingConfiguration'

-> RecordingConfiguration

Create a value of RecordingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:RecordingConfiguration', recordingConfiguration_name - Recording-configuration name. The value does not need to be unique.

$sel:tags:RecordingConfiguration', recordingConfiguration_tags - Array of 1-50 maps, each of the form string:string (key:value).

$sel:arn:RecordingConfiguration', recordingConfiguration_arn - Recording-configuration ARN.

$sel:destinationConfiguration:RecordingConfiguration', recordingConfiguration_destinationConfiguration - A complex type that contains information about where recorded video will be stored.

$sel:state:RecordingConfiguration', recordingConfiguration_state - Indicates the current state of the recording configuration. When the state is ACTIVE, the configuration is ready for recording a channel stream.

recordingConfiguration_name :: Lens' RecordingConfiguration (Maybe Text)

Recording-configuration name. The value does not need to be unique.

recordingConfiguration_tags :: Lens' RecordingConfiguration (Maybe (HashMap Text Text))

Array of 1-50 maps, each of the form string:string (key:value).

recordingConfiguration_arn :: Lens' RecordingConfiguration Text

Recording-configuration ARN.

recordingConfiguration_destinationConfiguration :: Lens' RecordingConfiguration DestinationConfiguration

A complex type that contains information about where recorded video will be stored.

recordingConfiguration_state :: Lens' RecordingConfiguration RecordingConfigurationState

Indicates the current state of the recording configuration. When the state is ACTIVE, the configuration is ready for recording a channel stream.