Network.AWS.IVS.Types.StreamSummary

Description

Documentation

data StreamSummary

Summary information about a stream.

See: newStreamSummary smart constructor.

Constructors

StreamSummary'

Fields

state :: Maybe StreamState

The stream’s state.

startTime :: Maybe POSIX

Time of the stream’s start. This is an ISO 8601 timestamp returned as a string.

channelArn :: Maybe Text

Channel ARN for the stream.

viewerCount :: Maybe Integer

A count of concurrent views of the stream. Typically, a new view appears in viewerCount within 15 seconds of when video playback starts and a view is removed from viewerCount within 1 minute of when video playback ends. A value of -1 indicates that the request timed out; in this case, retry.

health :: Maybe StreamHealth

The stream’s health.

newStreamSummary :: StreamSummary

Create a value of StreamSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:StreamSummary', streamSummary_state - The stream’s state.

$sel:startTime:StreamSummary', streamSummary_startTime - Time of the stream’s start. This is an ISO 8601 timestamp returned as a string.

$sel:channelArn:StreamSummary', streamSummary_channelArn - Channel ARN for the stream.

$sel:viewerCount:StreamSummary', streamSummary_viewerCount - A count of concurrent views of the stream. Typically, a new view appears in viewerCount within 15 seconds of when video playback starts and a view is removed from viewerCount within 1 minute of when video playback ends. A value of -1 indicates that the request timed out; in this case, retry.

$sel:health:StreamSummary', streamSummary_health - The stream’s health.

streamSummary_state :: Lens' StreamSummary (Maybe StreamState)

The stream’s state.

streamSummary_startTime :: Lens' StreamSummary (Maybe UTCTime)

Time of the stream’s start. This is an ISO 8601 timestamp returned as a string.

streamSummary_channelArn :: Lens' StreamSummary (Maybe Text)

Channel ARN for the stream.

streamSummary_viewerCount :: Lens' StreamSummary (Maybe Integer)

A count of concurrent views of the stream. Typically, a new view appears in viewerCount within 15 seconds of when video playback starts and a view is removed from viewerCount within 1 minute of when video playback ends. A value of -1 indicates that the request timed out; in this case, retry.

streamSummary_health :: Lens' StreamSummary (Maybe StreamHealth)

The stream’s health.