Network.AWS.IdentityStore.Lens

Description

Operations

DescribeGroup

describeGroup_identityStoreId :: Lens' DescribeGroup Text

The globally unique identifier for the identity store, such as d-1234567890. In this example, d- is a fixed prefix, and 1234567890 is a randomly generated string that contains number and lower case letters. This value is generated at the time that a new identity store is created.

describeGroup_groupId :: Lens' DescribeGroup Text

The identifier for a group in the identity store.

describeGroupResponse_httpStatus :: Lens' DescribeGroupResponse Int

The response's http status code.

describeGroupResponse_groupId :: Lens' DescribeGroupResponse Text

The identifier for a group in the identity store.

describeGroupResponse_displayName :: Lens' DescribeGroupResponse Text

Contains the group’s display name value. The length limit is 1,024 characters. This value can consist of letters, accented characters, symbols, numbers, punctuation, tab, new line, carriage return, space, and nonbreaking space in this attribute. The characters <>;:% are excluded. This value is specified at the time that the group is created and stored as an attribute of the group object in the identity store.

ListUsers

listUsers_filters :: Lens' ListUsers (Maybe [Filter])

A list of Filter objects, which is used in the ListUsers and ListGroups request.

listUsers_nextToken :: Lens' ListUsers (Maybe Text)

The pagination token used for the ListUsers and ListGroups API operations. This value is generated by the identity store service. It is returned in the API response if the total results are more than the size of one page. This token is also returned when it is used in the API request to search for the next page.

listUsers_maxResults :: Lens' ListUsers (Maybe Natural)

The maximum number of results to be returned per request. This parameter is used in the ListUsers and ListGroups request to specify how many results to return in one page. The length limit is 50 characters.

listUsers_identityStoreId :: Lens' ListUsers Text

The globally unique identifier for the identity store, such as d-1234567890. In this example, d- is a fixed prefix, and 1234567890 is a randomly generated string that contains number and lower case letters. This value is generated at the time that a new identity store is created.

listUsersResponse_nextToken :: Lens' ListUsersResponse (Maybe Text)

The pagination token used for the ListUsers and ListGroups API operations. This value is generated by the identity store service. It is returned in the API response if the total results are more than the size of one page. This token is also returned when it is used in the API request to search for the next page.

listUsersResponse_httpStatus :: Lens' ListUsersResponse Int

The response's http status code.

listUsersResponse_users :: Lens' ListUsersResponse [User]

A list of User objects in the identity store.

DescribeUser

describeUser_identityStoreId :: Lens' DescribeUser Text

The globally unique identifier for the identity store, such as d-1234567890. In this example, d- is a fixed prefix, and 1234567890 is a randomly generated string that contains number and lower case letters. This value is generated at the time that a new identity store is created.

describeUser_userId :: Lens' DescribeUser Text

The identifier for a user in the identity store.

describeUserResponse_httpStatus :: Lens' DescribeUserResponse Int

The response's http status code.

describeUserResponse_userName :: Lens' DescribeUserResponse Text

Contains the user’s user name value. The length limit is 128 characters. This value can consist of letters, accented characters, symbols, numbers, and punctuation. The characters <>;:% are excluded. This value is specified at the time the user is created and stored as an attribute of the user object in the identity store.

describeUserResponse_userId :: Lens' DescribeUserResponse Text

The identifier for a user in the identity store.

ListGroups

listGroups_filters :: Lens' ListGroups (Maybe [Filter])

A list of Filter objects, which is used in the ListUsers and ListGroups request.

listGroups_nextToken :: Lens' ListGroups (Maybe Text)

The pagination token used for the ListUsers and ListGroups API operations. This value is generated by the identity store service. It is returned in the API response if the total results are more than the size of one page. This token is also returned when it is used in the API request to search for the next page.

listGroups_maxResults :: Lens' ListGroups (Maybe Natural)

The maximum number of results to be returned per request. This parameter is used in the ListUsers and ListGroups request to specify how many results to return in one page. The length limit is 50 characters.

listGroups_identityStoreId :: Lens' ListGroups Text

The globally unique identifier for the identity store, such as d-1234567890. In this example, d- is a fixed prefix, and 1234567890 is a randomly generated string that contains number and lower case letters. This value is generated at the time that a new identity store is created.

listGroupsResponse_nextToken :: Lens' ListGroupsResponse (Maybe Text)

The pagination token used for the ListUsers and ListGroups API operations. This value is generated by the identity store service. It is returned in the API response if the total results are more than the size of one page. This token is also returned when it1 is used in the API request to search for the next page.

listGroupsResponse_httpStatus :: Lens' ListGroupsResponse Int

The response's http status code.

listGroupsResponse_groups :: Lens' ListGroupsResponse [Group]

A list of Group objects in the identity store.

Types

Filter

filter_attributePath :: Lens' Filter Text

The attribute path that is used to specify which attribute name to search. Length limit is 255 characters. For example, UserName is a valid attribute path for the ListUsers API, and DisplayName is a valid attribute path for the ListGroups API.

filter_attributeValue :: Lens' Filter Text

Represents the data for an attribute. Each attribute value is described as a name-value pair.

Group

group_groupId :: Lens' Group Text

The identifier for a group in the identity store.

group_displayName :: Lens' Group Text

Contains the group’s display name value. The length limit is 1,024 characters. This value can consist of letters, accented characters, symbols, numbers, punctuation, tab, new line, carriage return, space, and nonbreaking space in this attribute. The characters <>;:% are excluded. This value is specified at the time the group is created and stored as an attribute of the group object in the identity store.

User

user_userName :: Lens' User Text

Contains the user’s user name value. The length limit is 128 characters. This value can consist of letters, accented characters, symbols, numbers, and punctuation. The characters <>;:% are excluded. This value is specified at the time the user is created and stored as an attribute of the user object in the identity store.

user_userId :: Lens' User Text

The identifier for a user in the identity store.