Network.AWS.IdentityStore.Types

Description

Service Configuration

defaultService :: Service

API version 2020-06-15 of the Amazon SSO Identity Store SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The request failed because it contains a syntax error.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the principal has crossed the throttling limits of the API operations.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

The request processing has failed because of an unknown error, exception or failure with an internal server.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that a requested resource is not found.

Filter

data Filter

A query filter used by ListUsers and ListGroup. This filter object provides the attribute name and attribute value to search users or groups.

See: newFilter smart constructor.

Constructors

Filter'

Fields

attributePath :: Text

The attribute path that is used to specify which attribute name to search. Length limit is 255 characters. For example, UserName is a valid attribute path for the ListUsers API, and DisplayName is a valid attribute path for the ListGroups API.

attributeValue :: Sensitive Text

Represents the data for an attribute. Each attribute value is described as a name-value pair.

newFilter

Arguments

:: Text
$sel:attributePath:Filter'

-> Text
$sel:attributeValue:Filter'

-> Filter

Create a value of Filter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributePath:Filter', filter_attributePath - The attribute path that is used to specify which attribute name to search. Length limit is 255 characters. For example, UserName is a valid attribute path for the ListUsers API, and DisplayName is a valid attribute path for the ListGroups API.

$sel:attributeValue:Filter', filter_attributeValue - Represents the data for an attribute. Each attribute value is described as a name-value pair.

filter_attributePath :: Lens' Filter Text

The attribute path that is used to specify which attribute name to search. Length limit is 255 characters. For example, UserName is a valid attribute path for the ListUsers API, and DisplayName is a valid attribute path for the ListGroups API.

filter_attributeValue :: Lens' Filter Text

Represents the data for an attribute. Each attribute value is described as a name-value pair.

Group

data Group

A group object, which contains a specified group’s metadata and attributes.

See: newGroup smart constructor.

Constructors

Group'

Fields

groupId :: Text

The identifier for a group in the identity store.

displayName :: Text

Contains the group’s display name value. The length limit is 1,024 characters. This value can consist of letters, accented characters, symbols, numbers, punctuation, tab, new line, carriage return, space, and nonbreaking space in this attribute. The characters <>;:% are excluded. This value is specified at the time the group is created and stored as an attribute of the group object in the identity store.

newGroup

Arguments

:: Text
$sel:groupId:Group'

-> Text
$sel:displayName:Group'

-> Group

Create a value of Group with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupId:Group', group_groupId - The identifier for a group in the identity store.

$sel:displayName:Group', group_displayName - Contains the group’s display name value. The length limit is 1,024 characters. This value can consist of letters, accented characters, symbols, numbers, punctuation, tab, new line, carriage return, space, and nonbreaking space in this attribute. The characters <>;:% are excluded. This value is specified at the time the group is created and stored as an attribute of the group object in the identity store.

group_groupId :: Lens' Group Text

The identifier for a group in the identity store.

group_displayName :: Lens' Group Text

Contains the group’s display name value. The length limit is 1,024 characters. This value can consist of letters, accented characters, symbols, numbers, punctuation, tab, new line, carriage return, space, and nonbreaking space in this attribute. The characters <>;:% are excluded. This value is specified at the time the group is created and stored as an attribute of the group object in the identity store.

User

data User

A user object, which contains a specified user’s metadata and attributes.

See: newUser smart constructor.

Constructors

User'

Fields

userName :: Sensitive Text

Contains the user’s user name value. The length limit is 128 characters. This value can consist of letters, accented characters, symbols, numbers, and punctuation. The characters <>;:% are excluded. This value is specified at the time the user is created and stored as an attribute of the user object in the identity store.

userId :: Text

The identifier for a user in the identity store.

newUser

Arguments

:: Text
$sel:userName:User'

-> Text
$sel:userId:User'

-> User

Create a value of User with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userName:User', user_userName - Contains the user’s user name value. The length limit is 128 characters. This value can consist of letters, accented characters, symbols, numbers, and punctuation. The characters <>;:% are excluded. This value is specified at the time the user is created and stored as an attribute of the user object in the identity store.

$sel:userId:User', user_userId - The identifier for a user in the identity store.

user_userName :: Lens' User Text

Contains the user’s user name value. The length limit is 128 characters. This value can consist of letters, accented characters, symbols, numbers, and punctuation. The characters <>;:% are excluded. This value is specified at the time the user is created and stored as an attribute of the user object in the identity store.

user_userId :: Lens' User Text

The identifier for a user in the identity store.