Network.AWS.ImageBuilder.CancelImageCreation

Description

CancelImageCreation cancels the creation of Image. This operation can only be used on images in a non-terminal state.

Creating a Request

data CancelImageCreation

See: newCancelImageCreation smart constructor.

Constructors

CancelImageCreation'

Fields

imageBuildVersionArn :: Text

The Amazon Resource Name (ARN) of the image whose creation you want to cancel.

clientToken :: Text

The idempotency token used to make this request idempotent.

newCancelImageCreation

Arguments

:: Text
$sel:imageBuildVersionArn:CancelImageCreation'

-> Text
$sel:clientToken:CancelImageCreation'

-> CancelImageCreation

Create a value of CancelImageCreation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageBuildVersionArn:CancelImageCreation', cancelImageCreation_imageBuildVersionArn - The Amazon Resource Name (ARN) of the image whose creation you want to cancel.

$sel:clientToken:CancelImageCreation', cancelImageCreation_clientToken - The idempotency token used to make this request idempotent.

Request Lenses

cancelImageCreation_imageBuildVersionArn :: Lens' CancelImageCreation Text

The Amazon Resource Name (ARN) of the image whose creation you want to cancel.

cancelImageCreation_clientToken :: Lens' CancelImageCreation Text

The idempotency token used to make this request idempotent.

Destructuring the Response

data CancelImageCreationResponse

See: newCancelImageCreationResponse smart constructor.

Constructors

CancelImageCreationResponse'

Fields

requestId :: Maybe Text

The request ID that uniquely identifies this request.

clientToken :: Maybe Text

The idempotency token used to make this request idempotent.

imageBuildVersionArn :: Maybe Text

The Amazon Resource Name (ARN) of the image whose creation has been cancelled.

httpStatus :: Int

The response's http status code.

newCancelImageCreationResponse

Arguments

:: Int
$sel:httpStatus:CancelImageCreationResponse'

-> CancelImageCreationResponse

Create a value of CancelImageCreationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:CancelImageCreationResponse', cancelImageCreationResponse_requestId - The request ID that uniquely identifies this request.

$sel:clientToken:CancelImageCreation', cancelImageCreationResponse_clientToken - The idempotency token used to make this request idempotent.

$sel:imageBuildVersionArn:CancelImageCreation', cancelImageCreationResponse_imageBuildVersionArn - The Amazon Resource Name (ARN) of the image whose creation has been cancelled.

$sel:httpStatus:CancelImageCreationResponse', cancelImageCreationResponse_httpStatus - The response's http status code.

Response Lenses

cancelImageCreationResponse_requestId :: Lens' CancelImageCreationResponse (Maybe Text)

The request ID that uniquely identifies this request.

cancelImageCreationResponse_clientToken :: Lens' CancelImageCreationResponse (Maybe Text)

The idempotency token used to make this request idempotent.

cancelImageCreationResponse_imageBuildVersionArn :: Lens' CancelImageCreationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the image whose creation has been cancelled.

cancelImageCreationResponse_httpStatus :: Lens' CancelImageCreationResponse Int

The response's http status code.