Network.AWS.ImageBuilder.CreateContainerRecipe

Description

Creates a new container recipe. Container recipes define how images are configured, tested, and assessed.

Creating a Request

data CreateContainerRecipe

See: newCreateContainerRecipe smart constructor.

Constructors

CreateContainerRecipe'

Fields

dockerfileTemplateData :: Maybe Text

The Dockerfile template used to build your image as an inline data blob.

imageOsVersionOverride :: Maybe Text

Specifies the operating system version for the base image.

workingDirectory :: Maybe Text

The working directory for use during build and test workflows.

dockerfileTemplateUri :: Maybe Text

The Amazon S3 URI for the Dockerfile that will be used to build your container image.

kmsKeyId :: Maybe Text

Identifies which KMS key is used to encrypt the container image.

platformOverride :: Maybe Platform

Specifies the operating system platform when you use a custom base image.

description :: Maybe Text

The description of the container recipe.

tags :: Maybe (HashMap Text Text)

Tags that are attached to the container recipe.

instanceConfiguration :: Maybe InstanceConfiguration

A group of options that can be used to configure an instance for building and testing container images.

containerType :: ContainerType

The type of container to create.

name :: Text

The name of the container recipe.

semanticVersion :: Text

The semantic version of the container recipe. This version follows the semantic version syntax.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Assignment: For the first three nodes you can assign any positive integer value, including zero, with an upper limit of 2^30-1, or 1073741823 for each node. Image Builder automatically assigns the build number to the fourth node.

Patterns: You can use any numeric pattern that adheres to the assignment requirements for the nodes that you can assign. For example, you might choose a software version pattern, such as 1.0.0, or a date, such as 2021.01.01.

components :: NonEmpty ComponentConfiguration

Components for build and test that are included in the container recipe.

parentImage :: Text

The base image for the container recipe.

targetRepository :: TargetContainerRepository

The destination repository for the container image.

clientToken :: Text

The client token used to make this request idempotent.

newCreateContainerRecipe

Arguments

:: ContainerType
$sel:containerType:CreateContainerRecipe'

-> Text
$sel:name:CreateContainerRecipe'

-> Text
$sel:semanticVersion:CreateContainerRecipe'

-> NonEmpty ComponentConfiguration
$sel:components:CreateContainerRecipe'

-> Text
$sel:parentImage:CreateContainerRecipe'

-> TargetContainerRepository
$sel:targetRepository:CreateContainerRecipe'

-> Text
$sel:clientToken:CreateContainerRecipe'

-> CreateContainerRecipe

Create a value of CreateContainerRecipe with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dockerfileTemplateData:CreateContainerRecipe', createContainerRecipe_dockerfileTemplateData - The Dockerfile template used to build your image as an inline data blob.

$sel:imageOsVersionOverride:CreateContainerRecipe', createContainerRecipe_imageOsVersionOverride - Specifies the operating system version for the base image.

$sel:workingDirectory:CreateContainerRecipe', createContainerRecipe_workingDirectory - The working directory for use during build and test workflows.

$sel:dockerfileTemplateUri:CreateContainerRecipe', createContainerRecipe_dockerfileTemplateUri - The Amazon S3 URI for the Dockerfile that will be used to build your container image.

$sel:kmsKeyId:CreateContainerRecipe', createContainerRecipe_kmsKeyId - Identifies which KMS key is used to encrypt the container image.

$sel:platformOverride:CreateContainerRecipe', createContainerRecipe_platformOverride - Specifies the operating system platform when you use a custom base image.

$sel:description:CreateContainerRecipe', createContainerRecipe_description - The description of the container recipe.

$sel:tags:CreateContainerRecipe', createContainerRecipe_tags - Tags that are attached to the container recipe.

$sel:instanceConfiguration:CreateContainerRecipe', createContainerRecipe_instanceConfiguration - A group of options that can be used to configure an instance for building and testing container images.

$sel:containerType:CreateContainerRecipe', createContainerRecipe_containerType - The type of container to create.

$sel:name:CreateContainerRecipe', createContainerRecipe_name - The name of the container recipe.

$sel:semanticVersion:CreateContainerRecipe', createContainerRecipe_semanticVersion - The semantic version of the container recipe. This version follows the semantic version syntax.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Assignment: For the first three nodes you can assign any positive integer value, including zero, with an upper limit of 2^30-1, or 1073741823 for each node. Image Builder automatically assigns the build number to the fourth node.

Patterns: You can use any numeric pattern that adheres to the assignment requirements for the nodes that you can assign. For example, you might choose a software version pattern, such as 1.0.0, or a date, such as 2021.01.01.

$sel:components:CreateContainerRecipe', createContainerRecipe_components - Components for build and test that are included in the container recipe.

$sel:parentImage:CreateContainerRecipe', createContainerRecipe_parentImage - The base image for the container recipe.

$sel:targetRepository:CreateContainerRecipe', createContainerRecipe_targetRepository - The destination repository for the container image.

$sel:clientToken:CreateContainerRecipe', createContainerRecipe_clientToken - The client token used to make this request idempotent.

Request Lenses

createContainerRecipe_dockerfileTemplateData :: Lens' CreateContainerRecipe (Maybe Text)

The Dockerfile template used to build your image as an inline data blob.

createContainerRecipe_imageOsVersionOverride :: Lens' CreateContainerRecipe (Maybe Text)

Specifies the operating system version for the base image.

createContainerRecipe_workingDirectory :: Lens' CreateContainerRecipe (Maybe Text)

The working directory for use during build and test workflows.

createContainerRecipe_dockerfileTemplateUri :: Lens' CreateContainerRecipe (Maybe Text)

The Amazon S3 URI for the Dockerfile that will be used to build your container image.

createContainerRecipe_kmsKeyId :: Lens' CreateContainerRecipe (Maybe Text)

Identifies which KMS key is used to encrypt the container image.

createContainerRecipe_platformOverride :: Lens' CreateContainerRecipe (Maybe Platform)

Specifies the operating system platform when you use a custom base image.

createContainerRecipe_description :: Lens' CreateContainerRecipe (Maybe Text)

The description of the container recipe.

createContainerRecipe_tags :: Lens' CreateContainerRecipe (Maybe (HashMap Text Text))

Tags that are attached to the container recipe.

createContainerRecipe_instanceConfiguration :: Lens' CreateContainerRecipe (Maybe InstanceConfiguration)

A group of options that can be used to configure an instance for building and testing container images.

createContainerRecipe_containerType :: Lens' CreateContainerRecipe ContainerType

The type of container to create.

createContainerRecipe_name :: Lens' CreateContainerRecipe Text

The name of the container recipe.

createContainerRecipe_semanticVersion :: Lens' CreateContainerRecipe Text

The semantic version of the container recipe. This version follows the semantic version syntax.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Assignment: For the first three nodes you can assign any positive integer value, including zero, with an upper limit of 2^30-1, or 1073741823 for each node. Image Builder automatically assigns the build number to the fourth node.

Patterns: You can use any numeric pattern that adheres to the assignment requirements for the nodes that you can assign. For example, you might choose a software version pattern, such as 1.0.0, or a date, such as 2021.01.01.

createContainerRecipe_components :: Lens' CreateContainerRecipe (NonEmpty ComponentConfiguration)

Components for build and test that are included in the container recipe.

createContainerRecipe_parentImage :: Lens' CreateContainerRecipe Text

The base image for the container recipe.

createContainerRecipe_targetRepository :: Lens' CreateContainerRecipe TargetContainerRepository

The destination repository for the container image.

createContainerRecipe_clientToken :: Lens' CreateContainerRecipe Text

The client token used to make this request idempotent.

Destructuring the Response

data CreateContainerRecipeResponse

See: newCreateContainerRecipeResponse smart constructor.

Constructors

CreateContainerRecipeResponse'

Fields

requestId :: Maybe Text

The request ID that uniquely identifies this request.

clientToken :: Maybe Text

The client token used to make this request idempotent.

containerRecipeArn :: Maybe Text

Returns the Amazon Resource Name (ARN) of the container recipe that the request created.

httpStatus :: Int

The response's http status code.

newCreateContainerRecipeResponse

Arguments

:: Int
$sel:httpStatus:CreateContainerRecipeResponse'

-> CreateContainerRecipeResponse

Create a value of CreateContainerRecipeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:CreateContainerRecipeResponse', createContainerRecipeResponse_requestId - The request ID that uniquely identifies this request.

$sel:clientToken:CreateContainerRecipe', createContainerRecipeResponse_clientToken - The client token used to make this request idempotent.

$sel:containerRecipeArn:CreateContainerRecipeResponse', createContainerRecipeResponse_containerRecipeArn - Returns the Amazon Resource Name (ARN) of the container recipe that the request created.

$sel:httpStatus:CreateContainerRecipeResponse', createContainerRecipeResponse_httpStatus - The response's http status code.

Response Lenses

createContainerRecipeResponse_requestId :: Lens' CreateContainerRecipeResponse (Maybe Text)

The request ID that uniquely identifies this request.

createContainerRecipeResponse_clientToken :: Lens' CreateContainerRecipeResponse (Maybe Text)

The client token used to make this request idempotent.

createContainerRecipeResponse_containerRecipeArn :: Lens' CreateContainerRecipeResponse (Maybe Text)

Returns the Amazon Resource Name (ARN) of the container recipe that the request created.

createContainerRecipeResponse_httpStatus :: Lens' CreateContainerRecipeResponse Int

The response's http status code.