Network.AWS.ImageBuilder.CreateDistributionConfiguration

Description

Creates a new distribution configuration. Distribution configurations define and configure the outputs of your pipeline.

Creating a Request

data CreateDistributionConfiguration

See: newCreateDistributionConfiguration smart constructor.

Constructors

CreateDistributionConfiguration'

Fields

description :: Maybe Text

The description of the distribution configuration.

tags :: Maybe (HashMap Text Text)

The tags of the distribution configuration.

name :: Text

The name of the distribution configuration.

distributions :: [Distribution]

The distributions of the distribution configuration.

clientToken :: Text

The idempotency token of the distribution configuration.

newCreateDistributionConfiguration

Arguments

:: Text
$sel:name:CreateDistributionConfiguration'

-> Text
$sel:clientToken:CreateDistributionConfiguration'

-> CreateDistributionConfiguration

Create a value of CreateDistributionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateDistributionConfiguration', createDistributionConfiguration_description - The description of the distribution configuration.

$sel:tags:CreateDistributionConfiguration', createDistributionConfiguration_tags - The tags of the distribution configuration.

$sel:name:CreateDistributionConfiguration', createDistributionConfiguration_name - The name of the distribution configuration.

$sel:distributions:CreateDistributionConfiguration', createDistributionConfiguration_distributions - The distributions of the distribution configuration.

$sel:clientToken:CreateDistributionConfiguration', createDistributionConfiguration_clientToken - The idempotency token of the distribution configuration.

Request Lenses

createDistributionConfiguration_description :: Lens' CreateDistributionConfiguration (Maybe Text)

The description of the distribution configuration.

createDistributionConfiguration_tags :: Lens' CreateDistributionConfiguration (Maybe (HashMap Text Text))

The tags of the distribution configuration.

createDistributionConfiguration_name :: Lens' CreateDistributionConfiguration Text

The name of the distribution configuration.

createDistributionConfiguration_distributions :: Lens' CreateDistributionConfiguration [Distribution]

The distributions of the distribution configuration.

createDistributionConfiguration_clientToken :: Lens' CreateDistributionConfiguration Text

The idempotency token of the distribution configuration.

Destructuring the Response

data CreateDistributionConfigurationResponse

See: newCreateDistributionConfigurationResponse smart constructor.

Constructors

CreateDistributionConfigurationResponse'

Fields

requestId :: Maybe Text

The request ID that uniquely identifies this request.

clientToken :: Maybe Text

The idempotency token used to make this request idempotent.

distributionConfigurationArn :: Maybe Text

The Amazon Resource Name (ARN) of the distribution configuration that was created by this request.

httpStatus :: Int

The response's http status code.

newCreateDistributionConfigurationResponse

Arguments

:: Int
$sel:httpStatus:CreateDistributionConfigurationResponse'

-> CreateDistributionConfigurationResponse

Create a value of CreateDistributionConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:CreateDistributionConfigurationResponse', createDistributionConfigurationResponse_requestId - The request ID that uniquely identifies this request.

$sel:clientToken:CreateDistributionConfiguration', createDistributionConfigurationResponse_clientToken - The idempotency token used to make this request idempotent.

$sel:distributionConfigurationArn:CreateDistributionConfigurationResponse', createDistributionConfigurationResponse_distributionConfigurationArn - The Amazon Resource Name (ARN) of the distribution configuration that was created by this request.

$sel:httpStatus:CreateDistributionConfigurationResponse', createDistributionConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

createDistributionConfigurationResponse_requestId :: Lens' CreateDistributionConfigurationResponse (Maybe Text)

The request ID that uniquely identifies this request.

createDistributionConfigurationResponse_clientToken :: Lens' CreateDistributionConfigurationResponse (Maybe Text)

The idempotency token used to make this request idempotent.

createDistributionConfigurationResponse_distributionConfigurationArn :: Lens' CreateDistributionConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the distribution configuration that was created by this request.

createDistributionConfigurationResponse_httpStatus :: Lens' CreateDistributionConfigurationResponse Int

The response's http status code.