Network.AWS.ImageBuilder.CreateImageRecipe

Description

Creates a new image recipe. Image recipes define how images are configured, tested, and assessed.

Creating a Request

data CreateImageRecipe

See: newCreateImageRecipe smart constructor.

Constructors

CreateImageRecipe'

Fields

additionalInstanceConfiguration :: Maybe AdditionalInstanceConfiguration

Specify additional settings and launch scripts for your build instances.

workingDirectory :: Maybe Text

The working directory used during build and test workflows.

blockDeviceMappings :: Maybe [InstanceBlockDeviceMapping]

The block device mappings of the image recipe.

description :: Maybe Text

The description of the image recipe.

tags :: Maybe (HashMap Text Text)

The tags of the image recipe.

name :: Text

The name of the image recipe.

semanticVersion :: Text

The semantic version of the image recipe. This version follows the semantic version syntax.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Assignment: For the first three nodes you can assign any positive integer value, including zero, with an upper limit of 2^30-1, or 1073741823 for each node. Image Builder automatically assigns the build number to the fourth node.

Patterns: You can use any numeric pattern that adheres to the assignment requirements for the nodes that you can assign. For example, you might choose a software version pattern, such as 1.0.0, or a date, such as 2021.01.01.

components :: NonEmpty ComponentConfiguration

The components of the image recipe.

parentImage :: Text

The base image of the image recipe. The value of the string can be the ARN of the base image or an AMI ID. The format for the ARN follows this example: arn:aws:imagebuilder:us-west-2:aws:image/windows-server-2016-english-full-base-x86/x.x.x. You can provide the specific version that you want to use, or you can use a wildcard in all of the fields. If you enter an AMI ID for the string value, you must have access to the AMI, and the AMI must be in the same Region in which you are using Image Builder.

clientToken :: Text

The idempotency token used to make this request idempotent.

newCreateImageRecipe

Arguments

:: Text
$sel:name:CreateImageRecipe'

-> Text
$sel:semanticVersion:CreateImageRecipe'

-> NonEmpty ComponentConfiguration
$sel:components:CreateImageRecipe'

-> Text
$sel:parentImage:CreateImageRecipe'

-> Text
$sel:clientToken:CreateImageRecipe'

-> CreateImageRecipe

Create a value of CreateImageRecipe with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:additionalInstanceConfiguration:CreateImageRecipe', createImageRecipe_additionalInstanceConfiguration - Specify additional settings and launch scripts for your build instances.

$sel:workingDirectory:CreateImageRecipe', createImageRecipe_workingDirectory - The working directory used during build and test workflows.

$sel:blockDeviceMappings:CreateImageRecipe', createImageRecipe_blockDeviceMappings - The block device mappings of the image recipe.

$sel:description:CreateImageRecipe', createImageRecipe_description - The description of the image recipe.

$sel:tags:CreateImageRecipe', createImageRecipe_tags - The tags of the image recipe.

$sel:name:CreateImageRecipe', createImageRecipe_name - The name of the image recipe.

$sel:semanticVersion:CreateImageRecipe', createImageRecipe_semanticVersion - The semantic version of the image recipe. This version follows the semantic version syntax.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Assignment: For the first three nodes you can assign any positive integer value, including zero, with an upper limit of 2^30-1, or 1073741823 for each node. Image Builder automatically assigns the build number to the fourth node.

Patterns: You can use any numeric pattern that adheres to the assignment requirements for the nodes that you can assign. For example, you might choose a software version pattern, such as 1.0.0, or a date, such as 2021.01.01.

$sel:components:CreateImageRecipe', createImageRecipe_components - The components of the image recipe.

$sel:parentImage:CreateImageRecipe', createImageRecipe_parentImage - The base image of the image recipe. The value of the string can be the ARN of the base image or an AMI ID. The format for the ARN follows this example: arn:aws:imagebuilder:us-west-2:aws:image/windows-server-2016-english-full-base-x86/x.x.x. You can provide the specific version that you want to use, or you can use a wildcard in all of the fields. If you enter an AMI ID for the string value, you must have access to the AMI, and the AMI must be in the same Region in which you are using Image Builder.

$sel:clientToken:CreateImageRecipe', createImageRecipe_clientToken - The idempotency token used to make this request idempotent.

Request Lenses

createImageRecipe_additionalInstanceConfiguration :: Lens' CreateImageRecipe (Maybe AdditionalInstanceConfiguration)

Specify additional settings and launch scripts for your build instances.

createImageRecipe_workingDirectory :: Lens' CreateImageRecipe (Maybe Text)

The working directory used during build and test workflows.

createImageRecipe_blockDeviceMappings :: Lens' CreateImageRecipe (Maybe [InstanceBlockDeviceMapping])

The block device mappings of the image recipe.

createImageRecipe_description :: Lens' CreateImageRecipe (Maybe Text)

The description of the image recipe.

createImageRecipe_tags :: Lens' CreateImageRecipe (Maybe (HashMap Text Text))

The tags of the image recipe.

createImageRecipe_name :: Lens' CreateImageRecipe Text

The name of the image recipe.

createImageRecipe_semanticVersion :: Lens' CreateImageRecipe Text

The semantic version of the image recipe. This version follows the semantic version syntax.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Assignment: For the first three nodes you can assign any positive integer value, including zero, with an upper limit of 2^30-1, or 1073741823 for each node. Image Builder automatically assigns the build number to the fourth node.

Patterns: You can use any numeric pattern that adheres to the assignment requirements for the nodes that you can assign. For example, you might choose a software version pattern, such as 1.0.0, or a date, such as 2021.01.01.

createImageRecipe_components :: Lens' CreateImageRecipe (NonEmpty ComponentConfiguration)

The components of the image recipe.

createImageRecipe_parentImage :: Lens' CreateImageRecipe Text

The base image of the image recipe. The value of the string can be the ARN of the base image or an AMI ID. The format for the ARN follows this example: arn:aws:imagebuilder:us-west-2:aws:image/windows-server-2016-english-full-base-x86/x.x.x. You can provide the specific version that you want to use, or you can use a wildcard in all of the fields. If you enter an AMI ID for the string value, you must have access to the AMI, and the AMI must be in the same Region in which you are using Image Builder.

createImageRecipe_clientToken :: Lens' CreateImageRecipe Text

The idempotency token used to make this request idempotent.

Destructuring the Response

data CreateImageRecipeResponse

See: newCreateImageRecipeResponse smart constructor.

Constructors

CreateImageRecipeResponse'

Fields

requestId :: Maybe Text

The request ID that uniquely identifies this request.

clientToken :: Maybe Text

The idempotency token used to make this request idempotent.

imageRecipeArn :: Maybe Text

The Amazon Resource Name (ARN) of the image recipe that was created by this request.

httpStatus :: Int

The response's http status code.

newCreateImageRecipeResponse

Arguments

:: Int
$sel:httpStatus:CreateImageRecipeResponse'

-> CreateImageRecipeResponse

Create a value of CreateImageRecipeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:CreateImageRecipeResponse', createImageRecipeResponse_requestId - The request ID that uniquely identifies this request.

$sel:clientToken:CreateImageRecipe', createImageRecipeResponse_clientToken - The idempotency token used to make this request idempotent.

$sel:imageRecipeArn:CreateImageRecipeResponse', createImageRecipeResponse_imageRecipeArn - The Amazon Resource Name (ARN) of the image recipe that was created by this request.

$sel:httpStatus:CreateImageRecipeResponse', createImageRecipeResponse_httpStatus - The response's http status code.

Response Lenses

createImageRecipeResponse_requestId :: Lens' CreateImageRecipeResponse (Maybe Text)

The request ID that uniquely identifies this request.

createImageRecipeResponse_clientToken :: Lens' CreateImageRecipeResponse (Maybe Text)

The idempotency token used to make this request idempotent.

createImageRecipeResponse_imageRecipeArn :: Lens' CreateImageRecipeResponse (Maybe Text)

The Amazon Resource Name (ARN) of the image recipe that was created by this request.

createImageRecipeResponse_httpStatus :: Lens' CreateImageRecipeResponse Int

The response's http status code.