Network.AWS.ImageBuilder.CreateInfrastructureConfiguration

Description

Creates a new infrastructure configuration. An infrastructure configuration defines the environment in which your image will be built and tested.

Creating a Request

data CreateInfrastructureConfiguration

See: newCreateInfrastructureConfiguration smart constructor.

Constructors

CreateInfrastructureConfiguration'

Fields

securityGroupIds :: Maybe [Text]

The security group IDs to associate with the instance used to customize your Amazon EC2 AMI.

snsTopicArn :: Maybe Text

The SNS topic on which to send image build events.

instanceTypes :: Maybe [Text]

The instance types of the infrastructure configuration. You can specify one or more instance types to use for this build. The service will pick one of these instance types based on availability.

keyPair :: Maybe Text

The key pair of the infrastructure configuration. You can use this to log on to and debug the instance used to create your image.

resourceTags :: Maybe (HashMap Text Text)

The tags attached to the resource created by Image Builder.

subnetId :: Maybe Text

The subnet ID in which to place the instance used to customize your Amazon EC2 AMI.

instanceMetadataOptions :: Maybe InstanceMetadataOptions

The instance metadata options that you can set for the HTTP requests that pipeline builds use to launch EC2 build and test instances.

logging :: Maybe Logging

The logging configuration of the infrastructure configuration.

description :: Maybe Text

The description of the infrastructure configuration.

tags :: Maybe (HashMap Text Text)

The tags of the infrastructure configuration.

terminateInstanceOnFailure :: Maybe Bool

The terminate instance on failure setting of the infrastructure configuration. Set to false if you want Image Builder to retain the instance used to configure your AMI if the build or test phase of your workflow fails.

name :: Text

The name of the infrastructure configuration.

instanceProfileName :: Text

The instance profile to associate with the instance used to customize your Amazon EC2 AMI.

clientToken :: Text

The idempotency token used to make this request idempotent.

newCreateInfrastructureConfiguration

Arguments

:: Text
$sel:name:CreateInfrastructureConfiguration'

-> Text
$sel:instanceProfileName:CreateInfrastructureConfiguration'

-> Text
$sel:clientToken:CreateInfrastructureConfiguration'

-> CreateInfrastructureConfiguration

Create a value of CreateInfrastructureConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:CreateInfrastructureConfiguration', createInfrastructureConfiguration_securityGroupIds - The security group IDs to associate with the instance used to customize your Amazon EC2 AMI.

$sel:snsTopicArn:CreateInfrastructureConfiguration', createInfrastructureConfiguration_snsTopicArn - The SNS topic on which to send image build events.

$sel:instanceTypes:CreateInfrastructureConfiguration', createInfrastructureConfiguration_instanceTypes - The instance types of the infrastructure configuration. You can specify one or more instance types to use for this build. The service will pick one of these instance types based on availability.

$sel:keyPair:CreateInfrastructureConfiguration', createInfrastructureConfiguration_keyPair - The key pair of the infrastructure configuration. You can use this to log on to and debug the instance used to create your image.

$sel:resourceTags:CreateInfrastructureConfiguration', createInfrastructureConfiguration_resourceTags - The tags attached to the resource created by Image Builder.

$sel:subnetId:CreateInfrastructureConfiguration', createInfrastructureConfiguration_subnetId - The subnet ID in which to place the instance used to customize your Amazon EC2 AMI.

$sel:instanceMetadataOptions:CreateInfrastructureConfiguration', createInfrastructureConfiguration_instanceMetadataOptions - The instance metadata options that you can set for the HTTP requests that pipeline builds use to launch EC2 build and test instances.

$sel:logging:CreateInfrastructureConfiguration', createInfrastructureConfiguration_logging - The logging configuration of the infrastructure configuration.

$sel:description:CreateInfrastructureConfiguration', createInfrastructureConfiguration_description - The description of the infrastructure configuration.

$sel:tags:CreateInfrastructureConfiguration', createInfrastructureConfiguration_tags - The tags of the infrastructure configuration.

$sel:terminateInstanceOnFailure:CreateInfrastructureConfiguration', createInfrastructureConfiguration_terminateInstanceOnFailure - The terminate instance on failure setting of the infrastructure configuration. Set to false if you want Image Builder to retain the instance used to configure your AMI if the build or test phase of your workflow fails.

$sel:name:CreateInfrastructureConfiguration', createInfrastructureConfiguration_name - The name of the infrastructure configuration.

$sel:instanceProfileName:CreateInfrastructureConfiguration', createInfrastructureConfiguration_instanceProfileName - The instance profile to associate with the instance used to customize your Amazon EC2 AMI.

$sel:clientToken:CreateInfrastructureConfiguration', createInfrastructureConfiguration_clientToken - The idempotency token used to make this request idempotent.

Request Lenses

createInfrastructureConfiguration_securityGroupIds :: Lens' CreateInfrastructureConfiguration (Maybe [Text])

The security group IDs to associate with the instance used to customize your Amazon EC2 AMI.

createInfrastructureConfiguration_snsTopicArn :: Lens' CreateInfrastructureConfiguration (Maybe Text)

The SNS topic on which to send image build events.

createInfrastructureConfiguration_instanceTypes :: Lens' CreateInfrastructureConfiguration (Maybe [Text])

The instance types of the infrastructure configuration. You can specify one or more instance types to use for this build. The service will pick one of these instance types based on availability.

createInfrastructureConfiguration_keyPair :: Lens' CreateInfrastructureConfiguration (Maybe Text)

The key pair of the infrastructure configuration. You can use this to log on to and debug the instance used to create your image.

createInfrastructureConfiguration_resourceTags :: Lens' CreateInfrastructureConfiguration (Maybe (HashMap Text Text))

The tags attached to the resource created by Image Builder.

createInfrastructureConfiguration_subnetId :: Lens' CreateInfrastructureConfiguration (Maybe Text)

The subnet ID in which to place the instance used to customize your Amazon EC2 AMI.

createInfrastructureConfiguration_instanceMetadataOptions :: Lens' CreateInfrastructureConfiguration (Maybe InstanceMetadataOptions)

The instance metadata options that you can set for the HTTP requests that pipeline builds use to launch EC2 build and test instances.

createInfrastructureConfiguration_logging :: Lens' CreateInfrastructureConfiguration (Maybe Logging)

The logging configuration of the infrastructure configuration.

createInfrastructureConfiguration_description :: Lens' CreateInfrastructureConfiguration (Maybe Text)

The description of the infrastructure configuration.

createInfrastructureConfiguration_tags :: Lens' CreateInfrastructureConfiguration (Maybe (HashMap Text Text))

The tags of the infrastructure configuration.

createInfrastructureConfiguration_terminateInstanceOnFailure :: Lens' CreateInfrastructureConfiguration (Maybe Bool)

The terminate instance on failure setting of the infrastructure configuration. Set to false if you want Image Builder to retain the instance used to configure your AMI if the build or test phase of your workflow fails.

createInfrastructureConfiguration_name :: Lens' CreateInfrastructureConfiguration Text

The name of the infrastructure configuration.

createInfrastructureConfiguration_instanceProfileName :: Lens' CreateInfrastructureConfiguration Text

The instance profile to associate with the instance used to customize your Amazon EC2 AMI.

createInfrastructureConfiguration_clientToken :: Lens' CreateInfrastructureConfiguration Text

The idempotency token used to make this request idempotent.

Destructuring the Response

data CreateInfrastructureConfigurationResponse

See: newCreateInfrastructureConfigurationResponse smart constructor.

Constructors

CreateInfrastructureConfigurationResponse'

Fields

requestId :: Maybe Text

The request ID that uniquely identifies this request.

infrastructureConfigurationArn :: Maybe Text

The Amazon Resource Name (ARN) of the infrastructure configuration that was created by this request.

clientToken :: Maybe Text

The idempotency token used to make this request idempotent.

httpStatus :: Int

The response's http status code.

newCreateInfrastructureConfigurationResponse

Arguments

:: Int
$sel:httpStatus:CreateInfrastructureConfigurationResponse'

-> CreateInfrastructureConfigurationResponse

Create a value of CreateInfrastructureConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:CreateInfrastructureConfigurationResponse', createInfrastructureConfigurationResponse_requestId - The request ID that uniquely identifies this request.

$sel:infrastructureConfigurationArn:CreateInfrastructureConfigurationResponse', createInfrastructureConfigurationResponse_infrastructureConfigurationArn - The Amazon Resource Name (ARN) of the infrastructure configuration that was created by this request.

$sel:clientToken:CreateInfrastructureConfiguration', createInfrastructureConfigurationResponse_clientToken - The idempotency token used to make this request idempotent.

$sel:httpStatus:CreateInfrastructureConfigurationResponse', createInfrastructureConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

createInfrastructureConfigurationResponse_requestId :: Lens' CreateInfrastructureConfigurationResponse (Maybe Text)

The request ID that uniquely identifies this request.

createInfrastructureConfigurationResponse_infrastructureConfigurationArn :: Lens' CreateInfrastructureConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the infrastructure configuration that was created by this request.

createInfrastructureConfigurationResponse_clientToken :: Lens' CreateInfrastructureConfigurationResponse (Maybe Text)

The idempotency token used to make this request idempotent.

createInfrastructureConfigurationResponse_httpStatus :: Lens' CreateInfrastructureConfigurationResponse Int

The response's http status code.