Network.AWS.ImageBuilder.GetComponentPolicy

Description

Gets a component policy.

Creating a Request

data GetComponentPolicy

See: newGetComponentPolicy smart constructor.

Constructors

GetComponentPolicy'

Fields

componentArn :: Text

The Amazon Resource Name (ARN) of the component whose policy you want to retrieve.

newGetComponentPolicy

Arguments

:: Text
$sel:componentArn:GetComponentPolicy'

-> GetComponentPolicy

Create a value of GetComponentPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:componentArn:GetComponentPolicy', getComponentPolicy_componentArn - The Amazon Resource Name (ARN) of the component whose policy you want to retrieve.

Request Lenses

getComponentPolicy_componentArn :: Lens' GetComponentPolicy Text

The Amazon Resource Name (ARN) of the component whose policy you want to retrieve.

Destructuring the Response

data GetComponentPolicyResponse

See: newGetComponentPolicyResponse smart constructor.

Constructors

GetComponentPolicyResponse'

Fields

requestId :: Maybe Text

The request ID that uniquely identifies this request.

policy :: Maybe Text

The component policy.

httpStatus :: Int

The response's http status code.

newGetComponentPolicyResponse

Arguments

:: Int
$sel:httpStatus:GetComponentPolicyResponse'

-> GetComponentPolicyResponse

Create a value of GetComponentPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:GetComponentPolicyResponse', getComponentPolicyResponse_requestId - The request ID that uniquely identifies this request.

$sel:policy:GetComponentPolicyResponse', getComponentPolicyResponse_policy - The component policy.

$sel:httpStatus:GetComponentPolicyResponse', getComponentPolicyResponse_httpStatus - The response's http status code.

Response Lenses

getComponentPolicyResponse_requestId :: Lens' GetComponentPolicyResponse (Maybe Text)

The request ID that uniquely identifies this request.

getComponentPolicyResponse_policy :: Lens' GetComponentPolicyResponse (Maybe Text)

The component policy.

getComponentPolicyResponse_httpStatus :: Lens' GetComponentPolicyResponse Int

The response's http status code.