Network.AWS.ImageBuilder.GetImagePipeline

Description

Gets an image pipeline.

Creating a Request

data GetImagePipeline

See: newGetImagePipeline smart constructor.

Constructors

GetImagePipeline'

Fields

imagePipelineArn :: Text

The Amazon Resource Name (ARN) of the image pipeline that you want to retrieve.

newGetImagePipeline

Arguments

:: Text
$sel:imagePipelineArn:GetImagePipeline'

-> GetImagePipeline

Create a value of GetImagePipeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imagePipelineArn:GetImagePipeline', getImagePipeline_imagePipelineArn - The Amazon Resource Name (ARN) of the image pipeline that you want to retrieve.

Request Lenses

getImagePipeline_imagePipelineArn :: Lens' GetImagePipeline Text

The Amazon Resource Name (ARN) of the image pipeline that you want to retrieve.

Destructuring the Response

data GetImagePipelineResponse

See: newGetImagePipelineResponse smart constructor.

Constructors

GetImagePipelineResponse'

Fields

requestId :: Maybe Text

The request ID that uniquely identifies this request.

imagePipeline :: Maybe ImagePipeline

The image pipeline object.

httpStatus :: Int

The response's http status code.

newGetImagePipelineResponse

Arguments

:: Int
$sel:httpStatus:GetImagePipelineResponse'

-> GetImagePipelineResponse

Create a value of GetImagePipelineResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:GetImagePipelineResponse', getImagePipelineResponse_requestId - The request ID that uniquely identifies this request.

$sel:imagePipeline:GetImagePipelineResponse', getImagePipelineResponse_imagePipeline - The image pipeline object.

$sel:httpStatus:GetImagePipelineResponse', getImagePipelineResponse_httpStatus - The response's http status code.

Response Lenses

getImagePipelineResponse_requestId :: Lens' GetImagePipelineResponse (Maybe Text)

The request ID that uniquely identifies this request.

getImagePipelineResponse_imagePipeline :: Lens' GetImagePipelineResponse (Maybe ImagePipeline)

The image pipeline object.

getImagePipelineResponse_httpStatus :: Lens' GetImagePipelineResponse Int

The response's http status code.