Network.AWS.ImageBuilder.ImportComponent

Description

Imports a component and transforms its data into a component document.

Creating a Request

data ImportComponent

See: newImportComponent smart constructor.

Constructors

ImportComponent'

Fields

data' :: Maybe Text

The data of the component. Used to specify the data inline. Either data or uri can be used to specify the data within the component.

uri :: Maybe Text

The uri of the component. Must be an Amazon S3 URL and the requester must have permission to access the Amazon S3 bucket. If you use Amazon S3, you can specify component content up to your service quota. Either data or uri can be used to specify the data within the component.

kmsKeyId :: Maybe Text

The ID of the KMS key that should be used to encrypt this component.

changeDescription :: Maybe Text

The change description of the component. Describes what change has been made in this version, or what makes this version different from other versions of this component.

description :: Maybe Text

The description of the component. Describes the contents of the component.

tags :: Maybe (HashMap Text Text)

The tags of the component.

name :: Text

The name of the component.

semanticVersion :: Text

The semantic version of the component. This version follows the semantic version syntax.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Filtering: With semantic versioning, you have the flexibility to use wildcards (x) to specify the most recent versions or nodes when selecting the base image or components for your recipe. When you use a wildcard in any node, all nodes to the right of the first wildcard must also be wildcards.

type' :: ComponentType

The type of the component denotes whether the component is used to build the image, or only to test it.

format :: ComponentFormat

The format of the resource that you want to import as a component.

platform :: Platform

The platform of the component.

clientToken :: Text

The idempotency token of the component.

newImportComponent

Arguments

:: Text
$sel:name:ImportComponent'

-> Text
$sel:semanticVersion:ImportComponent'

-> ComponentType
$sel:type':ImportComponent'

-> ComponentFormat
$sel:format:ImportComponent'

-> Platform
$sel:platform:ImportComponent'

-> Text
$sel:clientToken:ImportComponent'

-> ImportComponent

Create a value of ImportComponent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:data':ImportComponent', importComponent_data - The data of the component. Used to specify the data inline. Either data or uri can be used to specify the data within the component.

$sel:uri:ImportComponent', importComponent_uri - The uri of the component. Must be an Amazon S3 URL and the requester must have permission to access the Amazon S3 bucket. If you use Amazon S3, you can specify component content up to your service quota. Either data or uri can be used to specify the data within the component.

$sel:kmsKeyId:ImportComponent', importComponent_kmsKeyId - The ID of the KMS key that should be used to encrypt this component.

$sel:changeDescription:ImportComponent', importComponent_changeDescription - The change description of the component. Describes what change has been made in this version, or what makes this version different from other versions of this component.

$sel:description:ImportComponent', importComponent_description - The description of the component. Describes the contents of the component.

$sel:tags:ImportComponent', importComponent_tags - The tags of the component.

$sel:name:ImportComponent', importComponent_name - The name of the component.

$sel:semanticVersion:ImportComponent', importComponent_semanticVersion - The semantic version of the component. This version follows the semantic version syntax.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Filtering: With semantic versioning, you have the flexibility to use wildcards (x) to specify the most recent versions or nodes when selecting the base image or components for your recipe. When you use a wildcard in any node, all nodes to the right of the first wildcard must also be wildcards.

$sel:type':ImportComponent', importComponent_type - The type of the component denotes whether the component is used to build the image, or only to test it.

$sel:format:ImportComponent', importComponent_format - The format of the resource that you want to import as a component.

$sel:platform:ImportComponent', importComponent_platform - The platform of the component.

$sel:clientToken:ImportComponent', importComponent_clientToken - The idempotency token of the component.

Request Lenses

importComponent_data :: Lens' ImportComponent (Maybe Text)

The data of the component. Used to specify the data inline. Either data or uri can be used to specify the data within the component.

importComponent_uri :: Lens' ImportComponent (Maybe Text)

The uri of the component. Must be an Amazon S3 URL and the requester must have permission to access the Amazon S3 bucket. If you use Amazon S3, you can specify component content up to your service quota. Either data or uri can be used to specify the data within the component.

importComponent_kmsKeyId :: Lens' ImportComponent (Maybe Text)

The ID of the KMS key that should be used to encrypt this component.

importComponent_changeDescription :: Lens' ImportComponent (Maybe Text)

The change description of the component. Describes what change has been made in this version, or what makes this version different from other versions of this component.

importComponent_description :: Lens' ImportComponent (Maybe Text)

The description of the component. Describes the contents of the component.

importComponent_tags :: Lens' ImportComponent (Maybe (HashMap Text Text))

The tags of the component.

importComponent_name :: Lens' ImportComponent Text

The name of the component.

importComponent_semanticVersion :: Lens' ImportComponent Text

The semantic version of the component. This version follows the semantic version syntax.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Filtering: With semantic versioning, you have the flexibility to use wildcards (x) to specify the most recent versions or nodes when selecting the base image or components for your recipe. When you use a wildcard in any node, all nodes to the right of the first wildcard must also be wildcards.

importComponent_type :: Lens' ImportComponent ComponentType

The type of the component denotes whether the component is used to build the image, or only to test it.

importComponent_format :: Lens' ImportComponent ComponentFormat

The format of the resource that you want to import as a component.

importComponent_platform :: Lens' ImportComponent Platform

The platform of the component.

importComponent_clientToken :: Lens' ImportComponent Text

The idempotency token of the component.

Destructuring the Response

data ImportComponentResponse

See: newImportComponentResponse smart constructor.

Constructors

ImportComponentResponse'

Fields

requestId :: Maybe Text

The request ID that uniquely identifies this request.

clientToken :: Maybe Text

The idempotency token used to make this request idempotent.

componentBuildVersionArn :: Maybe Text

The Amazon Resource Name (ARN) of the imported component.

httpStatus :: Int

The response's http status code.

newImportComponentResponse

Arguments

:: Int
$sel:httpStatus:ImportComponentResponse'

-> ImportComponentResponse

Create a value of ImportComponentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:ImportComponentResponse', importComponentResponse_requestId - The request ID that uniquely identifies this request.

$sel:clientToken:ImportComponent', importComponentResponse_clientToken - The idempotency token used to make this request idempotent.

$sel:componentBuildVersionArn:ImportComponentResponse', importComponentResponse_componentBuildVersionArn - The Amazon Resource Name (ARN) of the imported component.

$sel:httpStatus:ImportComponentResponse', importComponentResponse_httpStatus - The response's http status code.

Response Lenses

importComponentResponse_requestId :: Lens' ImportComponentResponse (Maybe Text)

The request ID that uniquely identifies this request.

importComponentResponse_clientToken :: Lens' ImportComponentResponse (Maybe Text)

The idempotency token used to make this request idempotent.

importComponentResponse_componentBuildVersionArn :: Lens' ImportComponentResponse (Maybe Text)

The Amazon Resource Name (ARN) of the imported component.

importComponentResponse_httpStatus :: Lens' ImportComponentResponse Int

The response's http status code.