Network.AWS.ImageBuilder.ListComponents

Description

Returns the list of component build versions for the specified semantic version.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Filtering: With semantic versioning, you have the flexibility to use wildcards (x) to specify the most recent versions or nodes when selecting the base image or components for your recipe. When you use a wildcard in any node, all nodes to the right of the first wildcard must also be wildcards.

Creating a Request

data ListComponents

See: newListComponents smart constructor.

Constructors

ListComponents'

Fields

filters :: Maybe (NonEmpty Filter)

Use the following filters to streamline results:

description

name

platform

supportedOsVersion

type

version

owner :: Maybe Ownership

The owner defines which components you want to list. By default, this request will only show components owned by your account. You can use this field to specify if you want to view components owned by yourself, by Amazon, or those components that have been shared with you by other customers.

byName :: Maybe Bool

Returns the list of component build versions for the specified name.

nextToken :: Maybe Text

A token to specify where to start paginating. This is the NextToken from a previously truncated response.

maxResults :: Maybe Natural

The maximum items to return in a request.

newListComponents :: ListComponents

Create a value of ListComponents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListComponents', listComponents_filters - Use the following filters to streamline results:

description

name

platform

supportedOsVersion

type

version

$sel:owner:ListComponents', listComponents_owner - The owner defines which components you want to list. By default, this request will only show components owned by your account. You can use this field to specify if you want to view components owned by yourself, by Amazon, or those components that have been shared with you by other customers.

$sel:byName:ListComponents', listComponents_byName - Returns the list of component build versions for the specified name.

$sel:nextToken:ListComponents', listComponents_nextToken - A token to specify where to start paginating. This is the NextToken from a previously truncated response.

$sel:maxResults:ListComponents', listComponents_maxResults - The maximum items to return in a request.

Request Lenses

listComponents_filters :: Lens' ListComponents (Maybe (NonEmpty Filter))

Use the following filters to streamline results:

description

name

platform

supportedOsVersion

type

version

listComponents_owner :: Lens' ListComponents (Maybe Ownership)

The owner defines which components you want to list. By default, this request will only show components owned by your account. You can use this field to specify if you want to view components owned by yourself, by Amazon, or those components that have been shared with you by other customers.

listComponents_byName :: Lens' ListComponents (Maybe Bool)

Returns the list of component build versions for the specified name.

listComponents_nextToken :: Lens' ListComponents (Maybe Text)

A token to specify where to start paginating. This is the NextToken from a previously truncated response.

listComponents_maxResults :: Lens' ListComponents (Maybe Natural)

The maximum items to return in a request.

Destructuring the Response

data ListComponentsResponse

See: newListComponentsResponse smart constructor.

Constructors

ListComponentsResponse'

Fields

requestId :: Maybe Text

The request ID that uniquely identifies this request.

componentVersionList :: Maybe [ComponentVersion]

The list of component semantic versions.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

nextToken :: Maybe Text

The next token used for paginated responses. When this is not empty, there are additional elements that the service has not included in this request. Use this token with the next request to retrieve additional objects.

httpStatus :: Int

The response's http status code.

newListComponentsResponse

Arguments

:: Int
$sel:httpStatus:ListComponentsResponse'

-> ListComponentsResponse

Create a value of ListComponentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:ListComponentsResponse', listComponentsResponse_requestId - The request ID that uniquely identifies this request.

$sel:componentVersionList:ListComponentsResponse', listComponentsResponse_componentVersionList - The list of component semantic versions.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

$sel:nextToken:ListComponents', listComponentsResponse_nextToken - The next token used for paginated responses. When this is not empty, there are additional elements that the service has not included in this request. Use this token with the next request to retrieve additional objects.

$sel:httpStatus:ListComponentsResponse', listComponentsResponse_httpStatus - The response's http status code.

Response Lenses

listComponentsResponse_requestId :: Lens' ListComponentsResponse (Maybe Text)

The request ID that uniquely identifies this request.

listComponentsResponse_componentVersionList :: Lens' ListComponentsResponse (Maybe [ComponentVersion])

The list of component semantic versions.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

listComponentsResponse_nextToken :: Lens' ListComponentsResponse (Maybe Text)

The next token used for paginated responses. When this is not empty, there are additional elements that the service has not included in this request. Use this token with the next request to retrieve additional objects.

listComponentsResponse_httpStatus :: Lens' ListComponentsResponse Int

The response's http status code.