Network.AWS.ImageBuilder.Types.ComponentSummary

Description

Documentation

data ComponentSummary

A high-level summary of a component.

See: newComponentSummary smart constructor.

Constructors

ComponentSummary'

Fields

state :: Maybe ComponentState

Describes the current status of the component.

platform :: Maybe Platform

The platform of the component.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the component.

supportedOsVersions :: Maybe (NonEmpty Text)

The operating system (OS) version supported by the component. If the OS information is available, a prefix match is performed against the base image OS version during image recipe creation.

owner :: Maybe Text

The owner of the component.

dateCreated :: Maybe Text

The date that the component was created.

name :: Maybe Text

The name of the component.

version :: Maybe Text

The version of the component.

changeDescription :: Maybe Text

The change description of the component.

type' :: Maybe ComponentType

The type of the component denotes whether the component is used to build the image or only to test it.

description :: Maybe Text

The description of the component.

tags :: Maybe (HashMap Text Text)

The tags associated with the component.

newComponentSummary :: ComponentSummary

Create a value of ComponentSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:ComponentSummary', componentSummary_state - Describes the current status of the component.

$sel:platform:ComponentSummary', componentSummary_platform - The platform of the component.

$sel:arn:ComponentSummary', componentSummary_arn - The Amazon Resource Name (ARN) of the component.

$sel:supportedOsVersions:ComponentSummary', componentSummary_supportedOsVersions - The operating system (OS) version supported by the component. If the OS information is available, a prefix match is performed against the base image OS version during image recipe creation.

$sel:owner:ComponentSummary', componentSummary_owner - The owner of the component.

$sel:dateCreated:ComponentSummary', componentSummary_dateCreated - The date that the component was created.

$sel:name:ComponentSummary', componentSummary_name - The name of the component.

$sel:version:ComponentSummary', componentSummary_version - The version of the component.

$sel:changeDescription:ComponentSummary', componentSummary_changeDescription - The change description of the component.

$sel:type':ComponentSummary', componentSummary_type - The type of the component denotes whether the component is used to build the image or only to test it.

$sel:description:ComponentSummary', componentSummary_description - The description of the component.

$sel:tags:ComponentSummary', componentSummary_tags - The tags associated with the component.

componentSummary_state :: Lens' ComponentSummary (Maybe ComponentState)

Describes the current status of the component.

componentSummary_platform :: Lens' ComponentSummary (Maybe Platform)

The platform of the component.

componentSummary_arn :: Lens' ComponentSummary (Maybe Text)

The Amazon Resource Name (ARN) of the component.

componentSummary_supportedOsVersions :: Lens' ComponentSummary (Maybe (NonEmpty Text))

The operating system (OS) version supported by the component. If the OS information is available, a prefix match is performed against the base image OS version during image recipe creation.

componentSummary_owner :: Lens' ComponentSummary (Maybe Text)

The owner of the component.

componentSummary_dateCreated :: Lens' ComponentSummary (Maybe Text)

The date that the component was created.

componentSummary_name :: Lens' ComponentSummary (Maybe Text)

The name of the component.

componentSummary_version :: Lens' ComponentSummary (Maybe Text)

The version of the component.

componentSummary_changeDescription :: Lens' ComponentSummary (Maybe Text)

The change description of the component.

componentSummary_type :: Lens' ComponentSummary (Maybe ComponentType)

The type of the component denotes whether the component is used to build the image or only to test it.

componentSummary_description :: Lens' ComponentSummary (Maybe Text)

The description of the component.

componentSummary_tags :: Lens' ComponentSummary (Maybe (HashMap Text Text))

The tags associated with the component.