Network.AWS.ImageBuilder.Types.ContainerDistributionConfiguration

Description

Documentation

data ContainerDistributionConfiguration

Container distribution settings for encryption, licensing, and sharing in a specific Region.

See: newContainerDistributionConfiguration smart constructor.

Constructors

ContainerDistributionConfiguration'

Fields

containerTags :: Maybe [Text]

Tags that are attached to the container distribution configuration.

description :: Maybe Text

The description of the container distribution configuration.

targetRepository :: TargetContainerRepository

The destination repository for the container distribution configuration.

newContainerDistributionConfiguration

Arguments

:: TargetContainerRepository
$sel:targetRepository:ContainerDistributionConfiguration'

-> ContainerDistributionConfiguration

Create a value of ContainerDistributionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerTags:ContainerDistributionConfiguration', containerDistributionConfiguration_containerTags - Tags that are attached to the container distribution configuration.

$sel:description:ContainerDistributionConfiguration', containerDistributionConfiguration_description - The description of the container distribution configuration.

$sel:targetRepository:ContainerDistributionConfiguration', containerDistributionConfiguration_targetRepository - The destination repository for the container distribution configuration.

containerDistributionConfiguration_containerTags :: Lens' ContainerDistributionConfiguration (Maybe [Text])

Tags that are attached to the container distribution configuration.

containerDistributionConfiguration_description :: Lens' ContainerDistributionConfiguration (Maybe Text)

The description of the container distribution configuration.

containerDistributionConfiguration_targetRepository :: Lens' ContainerDistributionConfiguration TargetContainerRepository

The destination repository for the container distribution configuration.