Network.AWS.ImageBuilder.Types.ContainerRecipeSummary

Description

Documentation

data ContainerRecipeSummary

A summary of a container recipe

See: newContainerRecipeSummary smart constructor.

Constructors

ContainerRecipeSummary'

Fields

containerType :: Maybe ContainerType

Specifies the type of container, such as "Docker".

platform :: Maybe Platform

The system platform for the container, such as Windows or Linux.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the container recipe.

parentImage :: Maybe Text

The base image for the container recipe.

owner :: Maybe Text

The owner of the container recipe.

dateCreated :: Maybe Text

The date when this container recipe was created.

name :: Maybe Text

The name of the container recipe.

tags :: Maybe (HashMap Text Text)

Tags that are attached to the container recipe.

newContainerRecipeSummary :: ContainerRecipeSummary

Create a value of ContainerRecipeSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerType:ContainerRecipeSummary', containerRecipeSummary_containerType - Specifies the type of container, such as "Docker".

$sel:platform:ContainerRecipeSummary', containerRecipeSummary_platform - The system platform for the container, such as Windows or Linux.

$sel:arn:ContainerRecipeSummary', containerRecipeSummary_arn - The Amazon Resource Name (ARN) of the container recipe.

$sel:parentImage:ContainerRecipeSummary', containerRecipeSummary_parentImage - The base image for the container recipe.

$sel:owner:ContainerRecipeSummary', containerRecipeSummary_owner - The owner of the container recipe.

$sel:dateCreated:ContainerRecipeSummary', containerRecipeSummary_dateCreated - The date when this container recipe was created.

$sel:name:ContainerRecipeSummary', containerRecipeSummary_name - The name of the container recipe.

$sel:tags:ContainerRecipeSummary', containerRecipeSummary_tags - Tags that are attached to the container recipe.

containerRecipeSummary_containerType :: Lens' ContainerRecipeSummary (Maybe ContainerType)

Specifies the type of container, such as "Docker".

containerRecipeSummary_platform :: Lens' ContainerRecipeSummary (Maybe Platform)

The system platform for the container, such as Windows or Linux.

containerRecipeSummary_arn :: Lens' ContainerRecipeSummary (Maybe Text)

The Amazon Resource Name (ARN) of the container recipe.

containerRecipeSummary_parentImage :: Lens' ContainerRecipeSummary (Maybe Text)

The base image for the container recipe.

containerRecipeSummary_owner :: Lens' ContainerRecipeSummary (Maybe Text)

The owner of the container recipe.

containerRecipeSummary_dateCreated :: Lens' ContainerRecipeSummary (Maybe Text)

The date when this container recipe was created.

containerRecipeSummary_name :: Lens' ContainerRecipeSummary (Maybe Text)

The name of the container recipe.

containerRecipeSummary_tags :: Lens' ContainerRecipeSummary (Maybe (HashMap Text Text))

Tags that are attached to the container recipe.