Network.AWS.ImageBuilder.Types.DistributionConfiguration

Description

Documentation

data DistributionConfiguration

A distribution configuration.

See: newDistributionConfiguration smart constructor.

Constructors

DistributionConfiguration'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the distribution configuration.

dateUpdated :: Maybe Text

The date on which this distribution configuration was last updated.

dateCreated :: Maybe Text

The date on which this distribution configuration was created.

name :: Maybe Text

The name of the distribution configuration.

description :: Maybe Text

The description of the distribution configuration.

distributions :: Maybe [Distribution]

The distribution objects that apply Region-specific settings for the deployment of the image to targeted Regions.

tags :: Maybe (HashMap Text Text)

The tags of the distribution configuration.

timeoutMinutes :: Natural

The maximum duration in minutes for this distribution configuration.

newDistributionConfiguration

Arguments

:: Natural
$sel:timeoutMinutes:DistributionConfiguration'

-> DistributionConfiguration

Create a value of DistributionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:DistributionConfiguration', distributionConfiguration_arn - The Amazon Resource Name (ARN) of the distribution configuration.

$sel:dateUpdated:DistributionConfiguration', distributionConfiguration_dateUpdated - The date on which this distribution configuration was last updated.

$sel:dateCreated:DistributionConfiguration', distributionConfiguration_dateCreated - The date on which this distribution configuration was created.

$sel:name:DistributionConfiguration', distributionConfiguration_name - The name of the distribution configuration.

$sel:description:DistributionConfiguration', distributionConfiguration_description - The description of the distribution configuration.

$sel:distributions:DistributionConfiguration', distributionConfiguration_distributions - The distribution objects that apply Region-specific settings for the deployment of the image to targeted Regions.

$sel:tags:DistributionConfiguration', distributionConfiguration_tags - The tags of the distribution configuration.

$sel:timeoutMinutes:DistributionConfiguration', distributionConfiguration_timeoutMinutes - The maximum duration in minutes for this distribution configuration.

distributionConfiguration_arn :: Lens' DistributionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the distribution configuration.

distributionConfiguration_dateUpdated :: Lens' DistributionConfiguration (Maybe Text)

The date on which this distribution configuration was last updated.

distributionConfiguration_dateCreated :: Lens' DistributionConfiguration (Maybe Text)

The date on which this distribution configuration was created.

distributionConfiguration_name :: Lens' DistributionConfiguration (Maybe Text)

The name of the distribution configuration.

distributionConfiguration_description :: Lens' DistributionConfiguration (Maybe Text)

The description of the distribution configuration.

distributionConfiguration_distributions :: Lens' DistributionConfiguration (Maybe [Distribution])

The distribution objects that apply Region-specific settings for the deployment of the image to targeted Regions.

distributionConfiguration_tags :: Lens' DistributionConfiguration (Maybe (HashMap Text Text))

The tags of the distribution configuration.

distributionConfiguration_timeoutMinutes :: Lens' DistributionConfiguration Natural

The maximum duration in minutes for this distribution configuration.