Network.AWS.ImageBuilder.Types.DistributionConfigurationSummary

Description

Documentation

data DistributionConfigurationSummary

A high-level overview of a distribution configuration.

See: newDistributionConfigurationSummary smart constructor.

Constructors

DistributionConfigurationSummary'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the distribution configuration.

regions :: Maybe [Text]

A list of Regions where the container image is distributed to.

dateUpdated :: Maybe Text

The date on which the distribution configuration was updated.

dateCreated :: Maybe Text

The date on which the distribution configuration was created.

name :: Maybe Text

The name of the distribution configuration.

description :: Maybe Text

The description of the distribution configuration.

tags :: Maybe (HashMap Text Text)

The tags associated with the distribution configuration.

newDistributionConfigurationSummary :: DistributionConfigurationSummary

Create a value of DistributionConfigurationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:DistributionConfigurationSummary', distributionConfigurationSummary_arn - The Amazon Resource Name (ARN) of the distribution configuration.

$sel:regions:DistributionConfigurationSummary', distributionConfigurationSummary_regions - A list of Regions where the container image is distributed to.

$sel:dateUpdated:DistributionConfigurationSummary', distributionConfigurationSummary_dateUpdated - The date on which the distribution configuration was updated.

$sel:dateCreated:DistributionConfigurationSummary', distributionConfigurationSummary_dateCreated - The date on which the distribution configuration was created.

$sel:name:DistributionConfigurationSummary', distributionConfigurationSummary_name - The name of the distribution configuration.

$sel:description:DistributionConfigurationSummary', distributionConfigurationSummary_description - The description of the distribution configuration.

$sel:tags:DistributionConfigurationSummary', distributionConfigurationSummary_tags - The tags associated with the distribution configuration.

distributionConfigurationSummary_arn :: Lens' DistributionConfigurationSummary (Maybe Text)

The Amazon Resource Name (ARN) of the distribution configuration.

distributionConfigurationSummary_regions :: Lens' DistributionConfigurationSummary (Maybe [Text])

A list of Regions where the container image is distributed to.

distributionConfigurationSummary_dateUpdated :: Lens' DistributionConfigurationSummary (Maybe Text)

The date on which the distribution configuration was updated.

distributionConfigurationSummary_dateCreated :: Lens' DistributionConfigurationSummary (Maybe Text)

The date on which the distribution configuration was created.

distributionConfigurationSummary_name :: Lens' DistributionConfigurationSummary (Maybe Text)

The name of the distribution configuration.

distributionConfigurationSummary_description :: Lens' DistributionConfigurationSummary (Maybe Text)

The description of the distribution configuration.

distributionConfigurationSummary_tags :: Lens' DistributionConfigurationSummary (Maybe (HashMap Text Text))

The tags associated with the distribution configuration.