Network.AWS.ImageBuilder.Types.Image

Description

Documentation

data Image

An Image Builder image. You must specify exactly one recipe for the image – either a container recipe (containerRecipe), which creates a container image, or an image recipe (imageRecipe), which creates an AMI.

See: newImage smart constructor.

Constructors

Image'

Fields

imageRecipe :: Maybe ImageRecipe

The image recipe used when creating the image.

state :: Maybe ImageState

The state of the image.

platform :: Maybe Platform

The platform of the image.

imageTestsConfiguration :: Maybe ImageTestsConfiguration

The image tests configuration used when creating this image.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the image.

Semantic versioning is included in each object's Amazon Resource Name (ARN), at the level that applies to that object as follows:

Versionless ARNs and Name ARNs do not include specific values in any of the nodes. The nodes are either left off entirely, or they are specified as wildcards, for example: x.x.x.

Version ARNs have only the first three nodes: <major>.<minor>.<patch>

Build version ARNs have all four nodes, and point to a specific build for a specific version of an object.

enhancedImageMetadataEnabled :: Maybe Bool

Collects additional information about the image being created, including the operating system (OS) version and package list. This information is used to enhance the overall experience of using EC2 Image Builder. Enabled by default.

dateCreated :: Maybe Text

The date on which this image was created.

containerRecipe :: Maybe ContainerRecipe

The recipe that is used to create an Image Builder container image.

name :: Maybe Text

The name of the image.

infrastructureConfiguration :: Maybe InfrastructureConfiguration

The infrastructure used when creating this image.

version :: Maybe Text

The semantic version of the image.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Assignment: For the first three nodes you can assign any positive integer value, including zero, with an upper limit of 2^30-1, or 1073741823 for each node. Image Builder automatically assigns the build number to the fourth node.

Patterns: You can use any numeric pattern that adheres to the assignment requirements for the nodes that you can assign. For example, you might choose a software version pattern, such as 1.0.0, or a date, such as 2021.01.01.

Filtering: With semantic versioning, you have the flexibility to use wildcards (x) to specify the most recent versions or nodes when selecting the base image or components for your recipe. When you use a wildcard in any node, all nodes to the right of the first wildcard must also be wildcards.

distributionConfiguration :: Maybe DistributionConfiguration

The distribution configuration used when creating this image.

sourcePipelineArn :: Maybe Text

The Amazon Resource Name (ARN) of the image pipeline that created this image.

type' :: Maybe ImageType

Specifies whether this is an AMI or container image.

sourcePipelineName :: Maybe Text

The name of the image pipeline that created this image.

outputResources :: Maybe OutputResources

The output resources produced when creating this image.

osVersion :: Maybe Text

The operating system version of the instance. For example, Amazon Linux 2, Ubuntu 18, or Microsoft Windows Server 2019.

tags :: Maybe (HashMap Text Text)

The tags of the image.

newImage :: Image

Create a value of Image with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageRecipe:Image', image_imageRecipe - The image recipe used when creating the image.

$sel:state:Image', image_state - The state of the image.

$sel:platform:Image', image_platform - The platform of the image.

$sel:imageTestsConfiguration:Image', image_imageTestsConfiguration - The image tests configuration used when creating this image.

$sel:arn:Image', image_arn - The Amazon Resource Name (ARN) of the image.

Semantic versioning is included in each object's Amazon Resource Name (ARN), at the level that applies to that object as follows:

Versionless ARNs and Name ARNs do not include specific values in any of the nodes. The nodes are either left off entirely, or they are specified as wildcards, for example: x.x.x.

Version ARNs have only the first three nodes: <major>.<minor>.<patch>

Build version ARNs have all four nodes, and point to a specific build for a specific version of an object.

$sel:enhancedImageMetadataEnabled:Image', image_enhancedImageMetadataEnabled - Collects additional information about the image being created, including the operating system (OS) version and package list. This information is used to enhance the overall experience of using EC2 Image Builder. Enabled by default.

$sel:dateCreated:Image', image_dateCreated - The date on which this image was created.

$sel:containerRecipe:Image', image_containerRecipe - The recipe that is used to create an Image Builder container image.

$sel:name:Image', image_name - The name of the image.

$sel:infrastructureConfiguration:Image', image_infrastructureConfiguration - The infrastructure used when creating this image.

$sel:version:Image', image_version - The semantic version of the image.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Assignment: For the first three nodes you can assign any positive integer value, including zero, with an upper limit of 2^30-1, or 1073741823 for each node. Image Builder automatically assigns the build number to the fourth node.

Patterns: You can use any numeric pattern that adheres to the assignment requirements for the nodes that you can assign. For example, you might choose a software version pattern, such as 1.0.0, or a date, such as 2021.01.01.

Filtering: With semantic versioning, you have the flexibility to use wildcards (x) to specify the most recent versions or nodes when selecting the base image or components for your recipe. When you use a wildcard in any node, all nodes to the right of the first wildcard must also be wildcards.

$sel:distributionConfiguration:Image', image_distributionConfiguration - The distribution configuration used when creating this image.

$sel:sourcePipelineArn:Image', image_sourcePipelineArn - The Amazon Resource Name (ARN) of the image pipeline that created this image.

$sel:type':Image', image_type - Specifies whether this is an AMI or container image.

$sel:sourcePipelineName:Image', image_sourcePipelineName - The name of the image pipeline that created this image.

$sel:outputResources:Image', image_outputResources - The output resources produced when creating this image.

$sel:osVersion:Image', image_osVersion - The operating system version of the instance. For example, Amazon Linux 2, Ubuntu 18, or Microsoft Windows Server 2019.

$sel:tags:Image', image_tags - The tags of the image.

image_imageRecipe :: Lens' Image (Maybe ImageRecipe)

The image recipe used when creating the image.

image_state :: Lens' Image (Maybe ImageState)

The state of the image.

image_platform :: Lens' Image (Maybe Platform)

The platform of the image.

image_imageTestsConfiguration :: Lens' Image (Maybe ImageTestsConfiguration)

The image tests configuration used when creating this image.

image_arn :: Lens' Image (Maybe Text)

The Amazon Resource Name (ARN) of the image.

Semantic versioning is included in each object's Amazon Resource Name (ARN), at the level that applies to that object as follows:

Versionless ARNs and Name ARNs do not include specific values in any of the nodes. The nodes are either left off entirely, or they are specified as wildcards, for example: x.x.x.

Version ARNs have only the first three nodes: <major>.<minor>.<patch>

Build version ARNs have all four nodes, and point to a specific build for a specific version of an object.

image_enhancedImageMetadataEnabled :: Lens' Image (Maybe Bool)

Collects additional information about the image being created, including the operating system (OS) version and package list. This information is used to enhance the overall experience of using EC2 Image Builder. Enabled by default.

image_dateCreated :: Lens' Image (Maybe Text)

The date on which this image was created.

image_containerRecipe :: Lens' Image (Maybe ContainerRecipe)

The recipe that is used to create an Image Builder container image.

image_name :: Lens' Image (Maybe Text)

The name of the image.

image_infrastructureConfiguration :: Lens' Image (Maybe InfrastructureConfiguration)

The infrastructure used when creating this image.

image_version :: Lens' Image (Maybe Text)

The semantic version of the image.

The semantic version has four nodes: <major>.<minor>.<patch>/<build>. You can assign values for the first three, and can filter on all of them.

Assignment: For the first three nodes you can assign any positive integer value, including zero, with an upper limit of 2^30-1, or 1073741823 for each node. Image Builder automatically assigns the build number to the fourth node.

Patterns: You can use any numeric pattern that adheres to the assignment requirements for the nodes that you can assign. For example, you might choose a software version pattern, such as 1.0.0, or a date, such as 2021.01.01.

Filtering: With semantic versioning, you have the flexibility to use wildcards (x) to specify the most recent versions or nodes when selecting the base image or components for your recipe. When you use a wildcard in any node, all nodes to the right of the first wildcard must also be wildcards.

image_distributionConfiguration :: Lens' Image (Maybe DistributionConfiguration)

The distribution configuration used when creating this image.

image_sourcePipelineArn :: Lens' Image (Maybe Text)

The Amazon Resource Name (ARN) of the image pipeline that created this image.

image_type :: Lens' Image (Maybe ImageType)

Specifies whether this is an AMI or container image.

image_sourcePipelineName :: Lens' Image (Maybe Text)

The name of the image pipeline that created this image.

image_outputResources :: Lens' Image (Maybe OutputResources)

The output resources produced when creating this image.

image_osVersion :: Lens' Image (Maybe Text)

The operating system version of the instance. For example, Amazon Linux 2, Ubuntu 18, or Microsoft Windows Server 2019.

image_tags :: Lens' Image (Maybe (HashMap Text Text))

The tags of the image.