Network.AWS.ImageBuilder.Types.ImagePipeline

Description

Documentation

data ImagePipeline

Details of an image pipeline.

See: newImagePipeline smart constructor.

Constructors

ImagePipeline'

Fields

status :: Maybe PipelineStatus

The status of the image pipeline.

dateLastRun :: Maybe Text

The date on which this image pipeline was last run.

platform :: Maybe Platform

The platform of the image pipeline.

infrastructureConfigurationArn :: Maybe Text

The Amazon Resource Name (ARN) of the infrastructure configuration associated with this image pipeline.

containerRecipeArn :: Maybe Text

The Amazon Resource Name (ARN) of the container recipe that is used for this pipeline.

imageTestsConfiguration :: Maybe ImageTestsConfiguration

The image tests configuration of the image pipeline.

dateNextRun :: Maybe Text

The date on which this image pipeline will next be run.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the image pipeline.

schedule :: Maybe Schedule

The schedule of the image pipeline.

enhancedImageMetadataEnabled :: Maybe Bool

Collects additional information about the image being created, including the operating system (OS) version and package list. This information is used to enhance the overall experience of using EC2 Image Builder. Enabled by default.

dateUpdated :: Maybe Text

The date on which this image pipeline was last updated.

dateCreated :: Maybe Text

The date on which this image pipeline was created.

name :: Maybe Text

The name of the image pipeline.

distributionConfigurationArn :: Maybe Text

The Amazon Resource Name (ARN) of the distribution configuration associated with this image pipeline.

imageRecipeArn :: Maybe Text

The Amazon Resource Name (ARN) of the image recipe associated with this image pipeline.

description :: Maybe Text

The description of the image pipeline.

tags :: Maybe (HashMap Text Text)

The tags of this image pipeline.

newImagePipeline :: ImagePipeline

Create a value of ImagePipeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ImagePipeline', imagePipeline_status - The status of the image pipeline.

$sel:dateLastRun:ImagePipeline', imagePipeline_dateLastRun - The date on which this image pipeline was last run.

$sel:platform:ImagePipeline', imagePipeline_platform - The platform of the image pipeline.

$sel:infrastructureConfigurationArn:ImagePipeline', imagePipeline_infrastructureConfigurationArn - The Amazon Resource Name (ARN) of the infrastructure configuration associated with this image pipeline.

$sel:containerRecipeArn:ImagePipeline', imagePipeline_containerRecipeArn - The Amazon Resource Name (ARN) of the container recipe that is used for this pipeline.

$sel:imageTestsConfiguration:ImagePipeline', imagePipeline_imageTestsConfiguration - The image tests configuration of the image pipeline.

$sel:dateNextRun:ImagePipeline', imagePipeline_dateNextRun - The date on which this image pipeline will next be run.

$sel:arn:ImagePipeline', imagePipeline_arn - The Amazon Resource Name (ARN) of the image pipeline.

$sel:schedule:ImagePipeline', imagePipeline_schedule - The schedule of the image pipeline.

$sel:enhancedImageMetadataEnabled:ImagePipeline', imagePipeline_enhancedImageMetadataEnabled - Collects additional information about the image being created, including the operating system (OS) version and package list. This information is used to enhance the overall experience of using EC2 Image Builder. Enabled by default.

$sel:dateUpdated:ImagePipeline', imagePipeline_dateUpdated - The date on which this image pipeline was last updated.

$sel:dateCreated:ImagePipeline', imagePipeline_dateCreated - The date on which this image pipeline was created.

$sel:name:ImagePipeline', imagePipeline_name - The name of the image pipeline.

$sel:distributionConfigurationArn:ImagePipeline', imagePipeline_distributionConfigurationArn - The Amazon Resource Name (ARN) of the distribution configuration associated with this image pipeline.

$sel:imageRecipeArn:ImagePipeline', imagePipeline_imageRecipeArn - The Amazon Resource Name (ARN) of the image recipe associated with this image pipeline.

$sel:description:ImagePipeline', imagePipeline_description - The description of the image pipeline.

$sel:tags:ImagePipeline', imagePipeline_tags - The tags of this image pipeline.

imagePipeline_status :: Lens' ImagePipeline (Maybe PipelineStatus)

The status of the image pipeline.

imagePipeline_dateLastRun :: Lens' ImagePipeline (Maybe Text)

The date on which this image pipeline was last run.

imagePipeline_platform :: Lens' ImagePipeline (Maybe Platform)

The platform of the image pipeline.

imagePipeline_infrastructureConfigurationArn :: Lens' ImagePipeline (Maybe Text)

The Amazon Resource Name (ARN) of the infrastructure configuration associated with this image pipeline.

imagePipeline_containerRecipeArn :: Lens' ImagePipeline (Maybe Text)

The Amazon Resource Name (ARN) of the container recipe that is used for this pipeline.

imagePipeline_imageTestsConfiguration :: Lens' ImagePipeline (Maybe ImageTestsConfiguration)

The image tests configuration of the image pipeline.

imagePipeline_dateNextRun :: Lens' ImagePipeline (Maybe Text)

The date on which this image pipeline will next be run.

imagePipeline_arn :: Lens' ImagePipeline (Maybe Text)

The Amazon Resource Name (ARN) of the image pipeline.

imagePipeline_schedule :: Lens' ImagePipeline (Maybe Schedule)

The schedule of the image pipeline.

imagePipeline_enhancedImageMetadataEnabled :: Lens' ImagePipeline (Maybe Bool)

Collects additional information about the image being created, including the operating system (OS) version and package list. This information is used to enhance the overall experience of using EC2 Image Builder. Enabled by default.

imagePipeline_dateUpdated :: Lens' ImagePipeline (Maybe Text)

The date on which this image pipeline was last updated.

imagePipeline_dateCreated :: Lens' ImagePipeline (Maybe Text)

The date on which this image pipeline was created.

imagePipeline_name :: Lens' ImagePipeline (Maybe Text)

The name of the image pipeline.

imagePipeline_distributionConfigurationArn :: Lens' ImagePipeline (Maybe Text)

The Amazon Resource Name (ARN) of the distribution configuration associated with this image pipeline.

imagePipeline_imageRecipeArn :: Lens' ImagePipeline (Maybe Text)

The Amazon Resource Name (ARN) of the image recipe associated with this image pipeline.

imagePipeline_description :: Lens' ImagePipeline (Maybe Text)

The description of the image pipeline.

imagePipeline_tags :: Lens' ImagePipeline (Maybe (HashMap Text Text))

The tags of this image pipeline.