Network.AWS.ImageBuilder.Types.ImageRecipe

Description

Documentation

data ImageRecipe

An image recipe.

See: newImageRecipe smart constructor.

Constructors

ImageRecipe'

Fields

components :: Maybe (NonEmpty ComponentConfiguration)

The components of the image recipe.

platform :: Maybe Platform

The platform of the image recipe.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the image recipe.

additionalInstanceConfiguration :: Maybe AdditionalInstanceConfiguration

Before you create a new AMI, Image Builder launches temporary Amazon EC2 instances to build and test your image configuration. Instance configuration adds a layer of control over those instances. You can define settings and add scripts to run when an instance is launched from your AMI.

workingDirectory :: Maybe Text

The working directory to be used during build and test workflows.

parentImage :: Maybe Text

The base image of the image recipe.

owner :: Maybe Text

The owner of the image recipe.

dateCreated :: Maybe Text

The date on which this image recipe was created.

name :: Maybe Text

The name of the image recipe.

version :: Maybe Text

The version of the image recipe.

type' :: Maybe ImageType

Specifies which type of image is created by the recipe - an AMI or a container image.

blockDeviceMappings :: Maybe [InstanceBlockDeviceMapping]

The block device mappings to apply when creating images from this recipe.

description :: Maybe Text

The description of the image recipe.

tags :: Maybe (HashMap Text Text)

The tags of the image recipe.

newImageRecipe :: ImageRecipe

Create a value of ImageRecipe with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:components:ImageRecipe', imageRecipe_components - The components of the image recipe.

$sel:platform:ImageRecipe', imageRecipe_platform - The platform of the image recipe.

$sel:arn:ImageRecipe', imageRecipe_arn - The Amazon Resource Name (ARN) of the image recipe.

$sel:additionalInstanceConfiguration:ImageRecipe', imageRecipe_additionalInstanceConfiguration - Before you create a new AMI, Image Builder launches temporary Amazon EC2 instances to build and test your image configuration. Instance configuration adds a layer of control over those instances. You can define settings and add scripts to run when an instance is launched from your AMI.

$sel:workingDirectory:ImageRecipe', imageRecipe_workingDirectory - The working directory to be used during build and test workflows.

$sel:parentImage:ImageRecipe', imageRecipe_parentImage - The base image of the image recipe.

$sel:owner:ImageRecipe', imageRecipe_owner - The owner of the image recipe.

$sel:dateCreated:ImageRecipe', imageRecipe_dateCreated - The date on which this image recipe was created.

$sel:name:ImageRecipe', imageRecipe_name - The name of the image recipe.

$sel:version:ImageRecipe', imageRecipe_version - The version of the image recipe.

$sel:type':ImageRecipe', imageRecipe_type - Specifies which type of image is created by the recipe - an AMI or a container image.

$sel:blockDeviceMappings:ImageRecipe', imageRecipe_blockDeviceMappings - The block device mappings to apply when creating images from this recipe.

$sel:description:ImageRecipe', imageRecipe_description - The description of the image recipe.

$sel:tags:ImageRecipe', imageRecipe_tags - The tags of the image recipe.

imageRecipe_components :: Lens' ImageRecipe (Maybe (NonEmpty ComponentConfiguration))

The components of the image recipe.

imageRecipe_platform :: Lens' ImageRecipe (Maybe Platform)

The platform of the image recipe.

imageRecipe_arn :: Lens' ImageRecipe (Maybe Text)

The Amazon Resource Name (ARN) of the image recipe.

imageRecipe_additionalInstanceConfiguration :: Lens' ImageRecipe (Maybe AdditionalInstanceConfiguration)

Before you create a new AMI, Image Builder launches temporary Amazon EC2 instances to build and test your image configuration. Instance configuration adds a layer of control over those instances. You can define settings and add scripts to run when an instance is launched from your AMI.

imageRecipe_workingDirectory :: Lens' ImageRecipe (Maybe Text)

The working directory to be used during build and test workflows.

imageRecipe_parentImage :: Lens' ImageRecipe (Maybe Text)

The base image of the image recipe.

imageRecipe_owner :: Lens' ImageRecipe (Maybe Text)

The owner of the image recipe.

imageRecipe_dateCreated :: Lens' ImageRecipe (Maybe Text)

The date on which this image recipe was created.

imageRecipe_name :: Lens' ImageRecipe (Maybe Text)

The name of the image recipe.

imageRecipe_version :: Lens' ImageRecipe (Maybe Text)

The version of the image recipe.

imageRecipe_type :: Lens' ImageRecipe (Maybe ImageType)

Specifies which type of image is created by the recipe - an AMI or a container image.

imageRecipe_blockDeviceMappings :: Lens' ImageRecipe (Maybe [InstanceBlockDeviceMapping])

The block device mappings to apply when creating images from this recipe.

imageRecipe_description :: Lens' ImageRecipe (Maybe Text)

The description of the image recipe.

imageRecipe_tags :: Lens' ImageRecipe (Maybe (HashMap Text Text))

The tags of the image recipe.