Network.AWS.ImageBuilder.Types.InfrastructureConfiguration

Description

Documentation

data InfrastructureConfiguration

Details of the infrastructure configuration.

See: newInfrastructureConfiguration smart constructor.

Constructors

InfrastructureConfiguration'

Fields

securityGroupIds :: Maybe [Text]

The security group IDs of the infrastructure configuration.

snsTopicArn :: Maybe Text

The SNS topic Amazon Resource Name (ARN) of the infrastructure configuration.

instanceTypes :: Maybe [Text]

The instance types of the infrastructure configuration.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the infrastructure configuration.

keyPair :: Maybe Text

The Amazon EC2 key pair of the infrastructure configuration.

resourceTags :: Maybe (HashMap Text Text)

The tags attached to the resource created by Image Builder.

subnetId :: Maybe Text

The subnet ID of the infrastructure configuration.

dateUpdated :: Maybe Text

The date on which the infrastructure configuration was last updated.

dateCreated :: Maybe Text

The date on which the infrastructure configuration was created.

instanceMetadataOptions :: Maybe InstanceMetadataOptions

The instance metadata option settings for the infrastructure configuration.

name :: Maybe Text

The name of the infrastructure configuration.

instanceProfileName :: Maybe Text

The instance profile of the infrastructure configuration.

logging :: Maybe Logging

The logging configuration of the infrastructure configuration.

description :: Maybe Text

The description of the infrastructure configuration.

tags :: Maybe (HashMap Text Text)

The tags of the infrastructure configuration.

terminateInstanceOnFailure :: Maybe Bool

The terminate instance on failure configuration of the infrastructure configuration.

newInfrastructureConfiguration :: InfrastructureConfiguration

Create a value of InfrastructureConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:InfrastructureConfiguration', infrastructureConfiguration_securityGroupIds - The security group IDs of the infrastructure configuration.

$sel:snsTopicArn:InfrastructureConfiguration', infrastructureConfiguration_snsTopicArn - The SNS topic Amazon Resource Name (ARN) of the infrastructure configuration.

$sel:instanceTypes:InfrastructureConfiguration', infrastructureConfiguration_instanceTypes - The instance types of the infrastructure configuration.

$sel:arn:InfrastructureConfiguration', infrastructureConfiguration_arn - The Amazon Resource Name (ARN) of the infrastructure configuration.

$sel:keyPair:InfrastructureConfiguration', infrastructureConfiguration_keyPair - The Amazon EC2 key pair of the infrastructure configuration.

$sel:resourceTags:InfrastructureConfiguration', infrastructureConfiguration_resourceTags - The tags attached to the resource created by Image Builder.

$sel:subnetId:InfrastructureConfiguration', infrastructureConfiguration_subnetId - The subnet ID of the infrastructure configuration.

$sel:dateUpdated:InfrastructureConfiguration', infrastructureConfiguration_dateUpdated - The date on which the infrastructure configuration was last updated.

$sel:dateCreated:InfrastructureConfiguration', infrastructureConfiguration_dateCreated - The date on which the infrastructure configuration was created.

$sel:instanceMetadataOptions:InfrastructureConfiguration', infrastructureConfiguration_instanceMetadataOptions - The instance metadata option settings for the infrastructure configuration.

$sel:name:InfrastructureConfiguration', infrastructureConfiguration_name - The name of the infrastructure configuration.

$sel:instanceProfileName:InfrastructureConfiguration', infrastructureConfiguration_instanceProfileName - The instance profile of the infrastructure configuration.

$sel:logging:InfrastructureConfiguration', infrastructureConfiguration_logging - The logging configuration of the infrastructure configuration.

$sel:description:InfrastructureConfiguration', infrastructureConfiguration_description - The description of the infrastructure configuration.

$sel:tags:InfrastructureConfiguration', infrastructureConfiguration_tags - The tags of the infrastructure configuration.

$sel:terminateInstanceOnFailure:InfrastructureConfiguration', infrastructureConfiguration_terminateInstanceOnFailure - The terminate instance on failure configuration of the infrastructure configuration.

infrastructureConfiguration_securityGroupIds :: Lens' InfrastructureConfiguration (Maybe [Text])

The security group IDs of the infrastructure configuration.

infrastructureConfiguration_snsTopicArn :: Lens' InfrastructureConfiguration (Maybe Text)

The SNS topic Amazon Resource Name (ARN) of the infrastructure configuration.

infrastructureConfiguration_instanceTypes :: Lens' InfrastructureConfiguration (Maybe [Text])

The instance types of the infrastructure configuration.

infrastructureConfiguration_arn :: Lens' InfrastructureConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the infrastructure configuration.

infrastructureConfiguration_keyPair :: Lens' InfrastructureConfiguration (Maybe Text)

The Amazon EC2 key pair of the infrastructure configuration.

infrastructureConfiguration_resourceTags :: Lens' InfrastructureConfiguration (Maybe (HashMap Text Text))

The tags attached to the resource created by Image Builder.

infrastructureConfiguration_subnetId :: Lens' InfrastructureConfiguration (Maybe Text)

The subnet ID of the infrastructure configuration.

infrastructureConfiguration_dateUpdated :: Lens' InfrastructureConfiguration (Maybe Text)

The date on which the infrastructure configuration was last updated.

infrastructureConfiguration_dateCreated :: Lens' InfrastructureConfiguration (Maybe Text)

The date on which the infrastructure configuration was created.

infrastructureConfiguration_instanceMetadataOptions :: Lens' InfrastructureConfiguration (Maybe InstanceMetadataOptions)

The instance metadata option settings for the infrastructure configuration.

infrastructureConfiguration_name :: Lens' InfrastructureConfiguration (Maybe Text)

The name of the infrastructure configuration.

infrastructureConfiguration_instanceProfileName :: Lens' InfrastructureConfiguration (Maybe Text)

The instance profile of the infrastructure configuration.

infrastructureConfiguration_logging :: Lens' InfrastructureConfiguration (Maybe Logging)

The logging configuration of the infrastructure configuration.

infrastructureConfiguration_description :: Lens' InfrastructureConfiguration (Maybe Text)

The description of the infrastructure configuration.

infrastructureConfiguration_tags :: Lens' InfrastructureConfiguration (Maybe (HashMap Text Text))

The tags of the infrastructure configuration.

infrastructureConfiguration_terminateInstanceOnFailure :: Lens' InfrastructureConfiguration (Maybe Bool)

The terminate instance on failure configuration of the infrastructure configuration.