Network.AWS.ImageBuilder.Types.LaunchPermissionConfiguration

Description

Documentation

data LaunchPermissionConfiguration

Describes the configuration for a launch permission. The launch permission modification request is sent to the Amazon EC2 ModifyImageAttribute API on behalf of the user for each Region they have selected to distribute the AMI. To make an AMI public, set the launch permission authorized accounts to all. See the examples for making an AMI public at Amazon EC2 ModifyImageAttribute.

See: newLaunchPermissionConfiguration smart constructor.

Constructors

LaunchPermissionConfiguration'

Fields

userIds :: Maybe (NonEmpty Text)

The Amazon Web Services account ID.

userGroups :: Maybe [Text]

The name of the group.

newLaunchPermissionConfiguration :: LaunchPermissionConfiguration

Create a value of LaunchPermissionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userIds:LaunchPermissionConfiguration', launchPermissionConfiguration_userIds - The Amazon Web Services account ID.

$sel:userGroups:LaunchPermissionConfiguration', launchPermissionConfiguration_userGroups - The name of the group.

launchPermissionConfiguration_userIds :: Lens' LaunchPermissionConfiguration (Maybe (NonEmpty Text))

The Amazon Web Services account ID.

launchPermissionConfiguration_userGroups :: Lens' LaunchPermissionConfiguration (Maybe [Text])

The name of the group.