Network.AWS.ImageBuilder.UpdateDistributionConfiguration

Description

Updates a new distribution configuration. Distribution configurations define and configure the outputs of your pipeline.

Creating a Request

data UpdateDistributionConfiguration

See: newUpdateDistributionConfiguration smart constructor.

Constructors

UpdateDistributionConfiguration'

Fields

description :: Maybe Text

The description of the distribution configuration.

distributionConfigurationArn :: Text

The Amazon Resource Name (ARN) of the distribution configuration that you want to update.

distributions :: [Distribution]

The distributions of the distribution configuration.

clientToken :: Text

The idempotency token of the distribution configuration.

newUpdateDistributionConfiguration

Arguments

:: Text
$sel:distributionConfigurationArn:UpdateDistributionConfiguration'

-> Text
$sel:clientToken:UpdateDistributionConfiguration'

-> UpdateDistributionConfiguration

Create a value of UpdateDistributionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:UpdateDistributionConfiguration', updateDistributionConfiguration_description - The description of the distribution configuration.

$sel:distributionConfigurationArn:UpdateDistributionConfiguration', updateDistributionConfiguration_distributionConfigurationArn - The Amazon Resource Name (ARN) of the distribution configuration that you want to update.

$sel:distributions:UpdateDistributionConfiguration', updateDistributionConfiguration_distributions - The distributions of the distribution configuration.

$sel:clientToken:UpdateDistributionConfiguration', updateDistributionConfiguration_clientToken - The idempotency token of the distribution configuration.

Request Lenses

updateDistributionConfiguration_description :: Lens' UpdateDistributionConfiguration (Maybe Text)

The description of the distribution configuration.

updateDistributionConfiguration_distributionConfigurationArn :: Lens' UpdateDistributionConfiguration Text

The Amazon Resource Name (ARN) of the distribution configuration that you want to update.

updateDistributionConfiguration_distributions :: Lens' UpdateDistributionConfiguration [Distribution]

The distributions of the distribution configuration.

updateDistributionConfiguration_clientToken :: Lens' UpdateDistributionConfiguration Text

The idempotency token of the distribution configuration.

Destructuring the Response

data UpdateDistributionConfigurationResponse

See: newUpdateDistributionConfigurationResponse smart constructor.

Constructors

UpdateDistributionConfigurationResponse'

Fields

requestId :: Maybe Text

The request ID that uniquely identifies this request.

clientToken :: Maybe Text

The idempotency token used to make this request idempotent.

distributionConfigurationArn :: Maybe Text

The Amazon Resource Name (ARN) of the distribution configuration that was updated by this request.

httpStatus :: Int

The response's http status code.

newUpdateDistributionConfigurationResponse

Arguments

:: Int
$sel:httpStatus:UpdateDistributionConfigurationResponse'

-> UpdateDistributionConfigurationResponse

Create a value of UpdateDistributionConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:UpdateDistributionConfigurationResponse', updateDistributionConfigurationResponse_requestId - The request ID that uniquely identifies this request.

$sel:clientToken:UpdateDistributionConfiguration', updateDistributionConfigurationResponse_clientToken - The idempotency token used to make this request idempotent.

$sel:distributionConfigurationArn:UpdateDistributionConfiguration', updateDistributionConfigurationResponse_distributionConfigurationArn - The Amazon Resource Name (ARN) of the distribution configuration that was updated by this request.

$sel:httpStatus:UpdateDistributionConfigurationResponse', updateDistributionConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

updateDistributionConfigurationResponse_requestId :: Lens' UpdateDistributionConfigurationResponse (Maybe Text)

The request ID that uniquely identifies this request.

updateDistributionConfigurationResponse_clientToken :: Lens' UpdateDistributionConfigurationResponse (Maybe Text)

The idempotency token used to make this request idempotent.

updateDistributionConfigurationResponse_distributionConfigurationArn :: Lens' UpdateDistributionConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the distribution configuration that was updated by this request.

updateDistributionConfigurationResponse_httpStatus :: Lens' UpdateDistributionConfigurationResponse Int

The response's http status code.