Network.AWS.ImageBuilder.UpdateInfrastructureConfiguration

Description

Updates a new infrastructure configuration. An infrastructure configuration defines the environment in which your image will be built and tested.

Creating a Request

data UpdateInfrastructureConfiguration

See: newUpdateInfrastructureConfiguration smart constructor.

Constructors

UpdateInfrastructureConfiguration'

Fields

securityGroupIds :: Maybe [Text]

The security group IDs to associate with the instance used to customize your Amazon EC2 AMI.

snsTopicArn :: Maybe Text

The SNS topic on which to send image build events.

instanceTypes :: Maybe [Text]

The instance types of the infrastructure configuration. You can specify one or more instance types to use for this build. The service will pick one of these instance types based on availability.

keyPair :: Maybe Text

The key pair of the infrastructure configuration. You can use this to log on to and debug the instance used to create your image.

resourceTags :: Maybe (HashMap Text Text)

The tags attached to the resource created by Image Builder.

subnetId :: Maybe Text

The subnet ID to place the instance used to customize your Amazon EC2 AMI in.

instanceMetadataOptions :: Maybe InstanceMetadataOptions

The instance metadata options that you can set for the HTTP requests that pipeline builds use to launch EC2 build and test instances. For more information about instance metadata options, see one of the following links:

Configure the instance metadata options in the /Amazon EC2 User Guide/ for Linux instances.

Configure the instance metadata options in the /Amazon EC2 Windows Guide/ for Windows instances.

logging :: Maybe Logging

The logging configuration of the infrastructure configuration.

description :: Maybe Text

The description of the infrastructure configuration.

terminateInstanceOnFailure :: Maybe Bool

The terminate instance on failure setting of the infrastructure configuration. Set to false if you want Image Builder to retain the instance used to configure your AMI if the build or test phase of your workflow fails.

infrastructureConfigurationArn :: Text

The Amazon Resource Name (ARN) of the infrastructure configuration that you want to update.

instanceProfileName :: Text

The instance profile to associate with the instance used to customize your Amazon EC2 AMI.

clientToken :: Text

The idempotency token used to make this request idempotent.

newUpdateInfrastructureConfiguration

Arguments

:: Text
$sel:infrastructureConfigurationArn:UpdateInfrastructureConfiguration'

-> Text
$sel:instanceProfileName:UpdateInfrastructureConfiguration'

-> Text
$sel:clientToken:UpdateInfrastructureConfiguration'

-> UpdateInfrastructureConfiguration

Create a value of UpdateInfrastructureConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:UpdateInfrastructureConfiguration', updateInfrastructureConfiguration_securityGroupIds - The security group IDs to associate with the instance used to customize your Amazon EC2 AMI.

$sel:snsTopicArn:UpdateInfrastructureConfiguration', updateInfrastructureConfiguration_snsTopicArn - The SNS topic on which to send image build events.

$sel:instanceTypes:UpdateInfrastructureConfiguration', updateInfrastructureConfiguration_instanceTypes - The instance types of the infrastructure configuration. You can specify one or more instance types to use for this build. The service will pick one of these instance types based on availability.

$sel:keyPair:UpdateInfrastructureConfiguration', updateInfrastructureConfiguration_keyPair - The key pair of the infrastructure configuration. You can use this to log on to and debug the instance used to create your image.

$sel:resourceTags:UpdateInfrastructureConfiguration', updateInfrastructureConfiguration_resourceTags - The tags attached to the resource created by Image Builder.

$sel:subnetId:UpdateInfrastructureConfiguration', updateInfrastructureConfiguration_subnetId - The subnet ID to place the instance used to customize your Amazon EC2 AMI in.

$sel:instanceMetadataOptions:UpdateInfrastructureConfiguration', updateInfrastructureConfiguration_instanceMetadataOptions - The instance metadata options that you can set for the HTTP requests that pipeline builds use to launch EC2 build and test instances. For more information about instance metadata options, see one of the following links:

Configure the instance metadata options in the /Amazon EC2 User Guide/ for Linux instances.

Configure the instance metadata options in the /Amazon EC2 Windows Guide/ for Windows instances.

$sel:logging:UpdateInfrastructureConfiguration', updateInfrastructureConfiguration_logging - The logging configuration of the infrastructure configuration.

$sel:description:UpdateInfrastructureConfiguration', updateInfrastructureConfiguration_description - The description of the infrastructure configuration.

$sel:terminateInstanceOnFailure:UpdateInfrastructureConfiguration', updateInfrastructureConfiguration_terminateInstanceOnFailure - The terminate instance on failure setting of the infrastructure configuration. Set to false if you want Image Builder to retain the instance used to configure your AMI if the build or test phase of your workflow fails.

$sel:infrastructureConfigurationArn:UpdateInfrastructureConfiguration', updateInfrastructureConfiguration_infrastructureConfigurationArn - The Amazon Resource Name (ARN) of the infrastructure configuration that you want to update.

$sel:instanceProfileName:UpdateInfrastructureConfiguration', updateInfrastructureConfiguration_instanceProfileName - The instance profile to associate with the instance used to customize your Amazon EC2 AMI.

$sel:clientToken:UpdateInfrastructureConfiguration', updateInfrastructureConfiguration_clientToken - The idempotency token used to make this request idempotent.

Request Lenses

updateInfrastructureConfiguration_securityGroupIds :: Lens' UpdateInfrastructureConfiguration (Maybe [Text])

The security group IDs to associate with the instance used to customize your Amazon EC2 AMI.

updateInfrastructureConfiguration_snsTopicArn :: Lens' UpdateInfrastructureConfiguration (Maybe Text)

The SNS topic on which to send image build events.

updateInfrastructureConfiguration_instanceTypes :: Lens' UpdateInfrastructureConfiguration (Maybe [Text])

The instance types of the infrastructure configuration. You can specify one or more instance types to use for this build. The service will pick one of these instance types based on availability.

updateInfrastructureConfiguration_keyPair :: Lens' UpdateInfrastructureConfiguration (Maybe Text)

The key pair of the infrastructure configuration. You can use this to log on to and debug the instance used to create your image.

updateInfrastructureConfiguration_resourceTags :: Lens' UpdateInfrastructureConfiguration (Maybe (HashMap Text Text))

The tags attached to the resource created by Image Builder.

updateInfrastructureConfiguration_subnetId :: Lens' UpdateInfrastructureConfiguration (Maybe Text)

The subnet ID to place the instance used to customize your Amazon EC2 AMI in.

updateInfrastructureConfiguration_instanceMetadataOptions :: Lens' UpdateInfrastructureConfiguration (Maybe InstanceMetadataOptions)

The instance metadata options that you can set for the HTTP requests that pipeline builds use to launch EC2 build and test instances. For more information about instance metadata options, see one of the following links:

Configure the instance metadata options in the /Amazon EC2 User Guide/ for Linux instances.

Configure the instance metadata options in the /Amazon EC2 Windows Guide/ for Windows instances.

updateInfrastructureConfiguration_logging :: Lens' UpdateInfrastructureConfiguration (Maybe Logging)

The logging configuration of the infrastructure configuration.

updateInfrastructureConfiguration_description :: Lens' UpdateInfrastructureConfiguration (Maybe Text)

The description of the infrastructure configuration.

updateInfrastructureConfiguration_terminateInstanceOnFailure :: Lens' UpdateInfrastructureConfiguration (Maybe Bool)

The terminate instance on failure setting of the infrastructure configuration. Set to false if you want Image Builder to retain the instance used to configure your AMI if the build or test phase of your workflow fails.

updateInfrastructureConfiguration_infrastructureConfigurationArn :: Lens' UpdateInfrastructureConfiguration Text

The Amazon Resource Name (ARN) of the infrastructure configuration that you want to update.

updateInfrastructureConfiguration_instanceProfileName :: Lens' UpdateInfrastructureConfiguration Text

The instance profile to associate with the instance used to customize your Amazon EC2 AMI.

updateInfrastructureConfiguration_clientToken :: Lens' UpdateInfrastructureConfiguration Text

The idempotency token used to make this request idempotent.

Destructuring the Response

data UpdateInfrastructureConfigurationResponse

See: newUpdateInfrastructureConfigurationResponse smart constructor.

Constructors

UpdateInfrastructureConfigurationResponse'

Fields

requestId :: Maybe Text

The request ID that uniquely identifies this request.

infrastructureConfigurationArn :: Maybe Text

The Amazon Resource Name (ARN) of the infrastructure configuration that was updated by this request.

clientToken :: Maybe Text

The idempotency token used to make this request idempotent.

httpStatus :: Int

The response's http status code.

newUpdateInfrastructureConfigurationResponse

Arguments

:: Int
$sel:httpStatus:UpdateInfrastructureConfigurationResponse'

-> UpdateInfrastructureConfigurationResponse

Create a value of UpdateInfrastructureConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:UpdateInfrastructureConfigurationResponse', updateInfrastructureConfigurationResponse_requestId - The request ID that uniquely identifies this request.

$sel:infrastructureConfigurationArn:UpdateInfrastructureConfiguration', updateInfrastructureConfigurationResponse_infrastructureConfigurationArn - The Amazon Resource Name (ARN) of the infrastructure configuration that was updated by this request.

$sel:clientToken:UpdateInfrastructureConfiguration', updateInfrastructureConfigurationResponse_clientToken - The idempotency token used to make this request idempotent.

$sel:httpStatus:UpdateInfrastructureConfigurationResponse', updateInfrastructureConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

updateInfrastructureConfigurationResponse_requestId :: Lens' UpdateInfrastructureConfigurationResponse (Maybe Text)

The request ID that uniquely identifies this request.

updateInfrastructureConfigurationResponse_infrastructureConfigurationArn :: Lens' UpdateInfrastructureConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the infrastructure configuration that was updated by this request.

updateInfrastructureConfigurationResponse_clientToken :: Lens' UpdateInfrastructureConfigurationResponse (Maybe Text)

The idempotency token used to make this request idempotent.

updateInfrastructureConfigurationResponse_httpStatus :: Lens' UpdateInfrastructureConfigurationResponse Int

The response's http status code.