Network.AWS.ImportExport

Description

Derived from API version 2010-06-01 of the AWS service descriptions, licensed under Apache 2.0.

AWS Import/Export Service AWS Import/Export accelerates transferring large amounts of data between the AWS cloud and portable storage devices that you mail to us. AWS Import/Export transfers data directly onto and off of your storage devices using Amazon's high-speed internal network and bypassing the Internet. For large data sets, AWS Import/Export is often faster than Internet transfer and more cost effective than upgrading your connectivity.

Service Configuration

defaultService :: Service

API version 2010-06-01 of the Amazon Import/Export SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by ImportExport.

InvalidJobIdException

_InvalidJobIdException :: AsError a => Getting (First ServiceError) a ServiceError

The JOBID was missing, not found, or not associated with the AWS account.

InvalidParameterException

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

One or more parameters had an invalid value.

ExpiredJobIdException

_ExpiredJobIdException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the specified job has expired out of the system.

InvalidFileSystemException

_InvalidFileSystemException :: AsError a => Getting (First ServiceError) a ServiceError

File system specified in export manifest is invalid.

InvalidAccessKeyIdException

_InvalidAccessKeyIdException :: AsError a => Getting (First ServiceError) a ServiceError

The AWS Access Key ID specified in the request did not match the manifest's accessKeyId value. The manifest and the request authentication must use the same AWS Access Key ID.

UnableToUpdateJobIdException

_UnableToUpdateJobIdException :: AsError a => Getting (First ServiceError) a ServiceError

AWS Import/Export cannot update the job

UnableToCancelJobIdException

_UnableToCancelJobIdException :: AsError a => Getting (First ServiceError) a ServiceError

AWS Import/Export cannot cancel the job

MultipleRegionsException

_MultipleRegionsException :: AsError a => Getting (First ServiceError) a ServiceError

Your manifest file contained buckets from multiple regions. A job is restricted to buckets from one region. Please correct and resubmit.

InvalidVersionException

_InvalidVersionException :: AsError a => Getting (First ServiceError) a ServiceError

The client tool version is invalid.

MalformedManifestException

_MalformedManifestException :: AsError a => Getting (First ServiceError) a ServiceError

Your manifest is not well-formed.

MissingParameterException

_MissingParameterException :: AsError a => Getting (First ServiceError) a ServiceError

One or more required parameters was missing from the request.

CanceledJobIdException

_CanceledJobIdException :: AsError a => Getting (First ServiceError) a ServiceError

The specified job ID has been canceled and is no longer valid.

BucketPermissionException

_BucketPermissionException :: AsError a => Getting (First ServiceError) a ServiceError

The account specified does not have the appropriate bucket permissions.

NoSuchBucketException

_NoSuchBucketException :: AsError a => Getting (First ServiceError) a ServiceError

The specified bucket does not exist. Create the specified bucket or change the manifest's bucket, exportBucket, or logBucket field to a bucket that the account, as specified by the manifest's Access Key ID, has write permissions to.

InvalidAddressException

_InvalidAddressException :: AsError a => Getting (First ServiceError) a ServiceError

The address specified in the manifest is invalid.

MissingCustomsException

_MissingCustomsException :: AsError a => Getting (First ServiceError) a ServiceError

One or more required customs parameters was missing from the manifest.

InvalidManifestFieldException

_InvalidManifestFieldException :: AsError a => Getting (First ServiceError) a ServiceError

One or more manifest fields was invalid. Please correct and resubmit.

InvalidCustomsException

_InvalidCustomsException :: AsError a => Getting (First ServiceError) a ServiceError

One or more customs parameters was invalid. Please correct and resubmit.

MissingManifestFieldException

_MissingManifestFieldException :: AsError a => Getting (First ServiceError) a ServiceError

One or more required fields were missing from the manifest file. Please correct and resubmit.

CreateJobQuotaExceededException

_CreateJobQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Each account can create only a certain number of jobs per day. If you need to create more than this, please contact awsimportexport@amazon.com to explain your particular use case.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

GetShippingLabel

data GetShippingLabel

See: newGetShippingLabel smart constructor.

Constructors

GetShippingLabel' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) [Text]

newGetShippingLabel :: GetShippingLabel

Create a value of GetShippingLabel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:street3:GetShippingLabel', getShippingLabel_street3 - Undocumented member.

$sel:aPIVersion:GetShippingLabel', getShippingLabel_aPIVersion - Undocumented member.

$sel:country:GetShippingLabel', getShippingLabel_country - Undocumented member.

$sel:stateOrProvince:GetShippingLabel', getShippingLabel_stateOrProvince - Undocumented member.

$sel:postalCode:GetShippingLabel', getShippingLabel_postalCode - Undocumented member.

$sel:street2:GetShippingLabel', getShippingLabel_street2 - Undocumented member.

$sel:name:GetShippingLabel', getShippingLabel_name - Undocumented member.

$sel:company:GetShippingLabel', getShippingLabel_company - Undocumented member.

$sel:phoneNumber:GetShippingLabel', getShippingLabel_phoneNumber - Undocumented member.

$sel:city:GetShippingLabel', getShippingLabel_city - Undocumented member.

$sel:street1:GetShippingLabel', getShippingLabel_street1 - Undocumented member.

$sel:jobIds:GetShippingLabel', getShippingLabel_jobIds - Undocumented member.

data GetShippingLabelResponse

See: newGetShippingLabelResponse smart constructor.

Constructors

GetShippingLabelResponse' (Maybe Text) (Maybe Text) Int

newGetShippingLabelResponse

Arguments

:: Int
$sel:httpStatus:GetShippingLabelResponse'

-> GetShippingLabelResponse

Create a value of GetShippingLabelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shippingLabelURL:GetShippingLabelResponse', getShippingLabelResponse_shippingLabelURL - Undocumented member.

$sel:warning:GetShippingLabelResponse', getShippingLabelResponse_warning - Undocumented member.

$sel:httpStatus:GetShippingLabelResponse', getShippingLabelResponse_httpStatus - The response's http status code.

CreateJob

data CreateJob

Input structure for the CreateJob operation.

See: newCreateJob smart constructor.

Constructors

CreateJob' (Maybe Text) (Maybe Text) JobType Text Bool

newCreateJob

Arguments

:: JobType
$sel:jobType:CreateJob'

-> Text
$sel:manifest:CreateJob'

-> Bool
$sel:validateOnly:CreateJob'

-> CreateJob

Create a value of CreateJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aPIVersion:CreateJob', createJob_aPIVersion - Undocumented member.

$sel:manifestAddendum:CreateJob', createJob_manifestAddendum - Undocumented member.

$sel:jobType:CreateJob', createJob_jobType - Undocumented member.

$sel:manifest:CreateJob', createJob_manifest - Undocumented member.

$sel:validateOnly:CreateJob', createJob_validateOnly - Undocumented member.

data CreateJobResponse

Output structure for the CreateJob operation.

See: newCreateJobResponse smart constructor.

Constructors

CreateJobResponse' (Maybe Text) (Maybe JobType) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe [Artifact]) Int

newCreateJobResponse

Arguments

:: Int
$sel:httpStatus:CreateJobResponse'

-> CreateJobResponse

Create a value of CreateJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:signature:CreateJobResponse', createJobResponse_signature - Undocumented member.

$sel:jobType:CreateJob', createJobResponse_jobType - Undocumented member.

$sel:jobId:CreateJobResponse', createJobResponse_jobId - Undocumented member.

$sel:signatureFileContents:CreateJobResponse', createJobResponse_signatureFileContents - Undocumented member.

$sel:warningMessage:CreateJobResponse', createJobResponse_warningMessage - Undocumented member.

$sel:artifactList:CreateJobResponse', createJobResponse_artifactList - Undocumented member.

$sel:httpStatus:CreateJobResponse', createJobResponse_httpStatus - The response's http status code.

ListJobs (Paginated)

data ListJobs

Input structure for the ListJobs operation.

See: newListJobs smart constructor.

Constructors

ListJobs' (Maybe Text) (Maybe Text) (Maybe Int)

newListJobs :: ListJobs

Create a value of ListJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aPIVersion:ListJobs', listJobs_aPIVersion - Undocumented member.

$sel:marker:ListJobs', listJobs_marker - Undocumented member.

$sel:maxJobs:ListJobs', listJobs_maxJobs - Undocumented member.

data ListJobsResponse

Output structure for the ListJobs operation.

See: newListJobsResponse smart constructor.

Constructors

ListJobsResponse' (Maybe [Job]) (Maybe Bool) Int

newListJobsResponse

Arguments

:: Int
$sel:httpStatus:ListJobsResponse'

-> ListJobsResponse

Create a value of ListJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobs:ListJobsResponse', listJobsResponse_jobs - Undocumented member.

$sel:isTruncated:ListJobsResponse', listJobsResponse_isTruncated - Undocumented member.

$sel:httpStatus:ListJobsResponse', listJobsResponse_httpStatus - The response's http status code.

UpdateJob

data UpdateJob

Input structure for the UpateJob operation.

See: newUpdateJob smart constructor.

Constructors

UpdateJob' (Maybe Text) Text Text JobType Bool

newUpdateJob

Arguments

:: Text
$sel:jobId:UpdateJob'

-> Text
$sel:manifest:UpdateJob'

-> JobType
$sel:jobType:UpdateJob'

-> Bool
$sel:validateOnly:UpdateJob'

-> UpdateJob

Create a value of UpdateJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aPIVersion:UpdateJob', updateJob_aPIVersion - Undocumented member.

$sel:jobId:UpdateJob', updateJob_jobId - Undocumented member.

$sel:manifest:UpdateJob', updateJob_manifest - Undocumented member.

$sel:jobType:UpdateJob', updateJob_jobType - Undocumented member.

$sel:validateOnly:UpdateJob', updateJob_validateOnly - Undocumented member.

data UpdateJobResponse

Output structure for the UpateJob operation.

See: newUpdateJobResponse smart constructor.

Constructors

UpdateJobResponse' (Maybe Bool) (Maybe Text) (Maybe [Artifact]) Int

newUpdateJobResponse

Arguments

:: Int
$sel:httpStatus:UpdateJobResponse'

-> UpdateJobResponse

Create a value of UpdateJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:success:UpdateJobResponse', updateJobResponse_success - Undocumented member.

$sel:warningMessage:UpdateJobResponse', updateJobResponse_warningMessage - Undocumented member.

$sel:artifactList:UpdateJobResponse', updateJobResponse_artifactList - Undocumented member.

$sel:httpStatus:UpdateJobResponse', updateJobResponse_httpStatus - The response's http status code.

GetStatus

data GetStatus

Input structure for the GetStatus operation.

See: newGetStatus smart constructor.

Constructors

GetStatus' (Maybe Text) Text

newGetStatus

Arguments

:: Text
$sel:jobId:GetStatus'

-> GetStatus

Create a value of GetStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aPIVersion:GetStatus', getStatus_aPIVersion - Undocumented member.

$sel:jobId:GetStatus', getStatus_jobId - Undocumented member.

data GetStatusResponse

Output structure for the GetStatus operation.

See: newGetStatusResponse smart constructor.

Constructors

GetStatusResponse' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe JobType) (Maybe Text) (Maybe Text) (Maybe Int) (Maybe Text) (Maybe [Artifact]) (Maybe Text) (Maybe ISO8601) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) Int

newGetStatusResponse

Arguments

:: Int
$sel:httpStatus:GetStatusResponse'

-> GetStatusResponse

Create a value of GetStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:carrier:GetStatusResponse', getStatusResponse_carrier - Undocumented member.

$sel:trackingNumber:GetStatusResponse', getStatusResponse_trackingNumber - Undocumented member.

$sel:signature:GetStatusResponse', getStatusResponse_signature - Undocumented member.

$sel:jobType:GetStatusResponse', getStatusResponse_jobType - Undocumented member.

$sel:jobId:GetStatus', getStatusResponse_jobId - Undocumented member.

$sel:signatureFileContents:GetStatusResponse', getStatusResponse_signatureFileContents - Undocumented member.

$sel:errorCount:GetStatusResponse', getStatusResponse_errorCount - Undocumented member.

$sel:currentManifest:GetStatusResponse', getStatusResponse_currentManifest - Undocumented member.

$sel:artifactList:GetStatusResponse', getStatusResponse_artifactList - Undocumented member.

$sel:logBucket:GetStatusResponse', getStatusResponse_logBucket - Undocumented member.

$sel:creationDate:GetStatusResponse', getStatusResponse_creationDate - Undocumented member.

$sel:progressCode:GetStatusResponse', getStatusResponse_progressCode - Undocumented member.

$sel:locationCode:GetStatusResponse', getStatusResponse_locationCode - Undocumented member.

$sel:logKey:GetStatusResponse', getStatusResponse_logKey - Undocumented member.

$sel:locationMessage:GetStatusResponse', getStatusResponse_locationMessage - Undocumented member.

$sel:progressMessage:GetStatusResponse', getStatusResponse_progressMessage - Undocumented member.

$sel:httpStatus:GetStatusResponse', getStatusResponse_httpStatus - The response's http status code.

CancelJob

data CancelJob

Input structure for the CancelJob operation.

See: newCancelJob smart constructor.

Constructors

CancelJob' (Maybe Text) Text

newCancelJob

Arguments

:: Text
$sel:jobId:CancelJob'

-> CancelJob

Create a value of CancelJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aPIVersion:CancelJob', cancelJob_aPIVersion - Undocumented member.

$sel:jobId:CancelJob', cancelJob_jobId - Undocumented member.

data CancelJobResponse

Output structure for the CancelJob operation.

See: newCancelJobResponse smart constructor.

Constructors

CancelJobResponse' (Maybe Bool) Int

newCancelJobResponse

Arguments

:: Int
$sel:httpStatus:CancelJobResponse'

-> CancelJobResponse

Create a value of CancelJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:success:CancelJobResponse', cancelJobResponse_success - Undocumented member.

$sel:httpStatus:CancelJobResponse', cancelJobResponse_httpStatus - The response's http status code.

Types

JobType

newtype JobType

Specifies whether the job to initiate is an import or export job.

Constructors

JobType'

Fields

fromJobType :: Text

Bundled Patterns

pattern JobType_Export :: JobType

pattern JobType_Import :: JobType

Artifact

data Artifact

A discrete item that contains the description and URL of an artifact (such as a PDF).

See: newArtifact smart constructor.

Constructors

Artifact' (Maybe Text) (Maybe Text)

newArtifact :: Artifact

Create a value of Artifact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:url:Artifact', artifact_url - Undocumented member.

$sel:description:Artifact', artifact_description - Undocumented member.

Job

data Job

Representation of a job returned by the ListJobs operation.

See: newJob smart constructor.

Constructors

Job' JobType Text Bool ISO8601

newJob

Arguments

:: JobType
$sel:jobType:Job'

-> Text
$sel:jobId:Job'

-> Bool
$sel:isCanceled:Job'

-> UTCTime
$sel:creationDate:Job'

-> Job

Create a value of Job with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobType:Job', job_jobType - Undocumented member.

$sel:jobId:Job', job_jobId - Undocumented member.

$sel:isCanceled:Job', job_isCanceled - Undocumented member.

$sel:creationDate:Job', job_creationDate - Undocumented member.