Network.AWS.ImportExport.GetStatus

Description

This operation returns information about a job, including where the job is in the processing pipeline, the status of the results, and the signature value associated with the job. You can only return information about jobs you own.

Creating a Request

data GetStatus

Input structure for the GetStatus operation.

See: newGetStatus smart constructor.

Constructors

GetStatus'

Fields

aPIVersion :: Maybe Text

jobId :: Text

newGetStatus

Arguments

:: Text
$sel:jobId:GetStatus'

-> GetStatus

Create a value of GetStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aPIVersion:GetStatus', getStatus_aPIVersion - Undocumented member.

$sel:jobId:GetStatus', getStatus_jobId - Undocumented member.

Request Lenses

getStatus_aPIVersion :: Lens' GetStatus (Maybe Text)

Undocumented member.

getStatus_jobId :: Lens' GetStatus Text

Undocumented member.

Destructuring the Response

data GetStatusResponse

Output structure for the GetStatus operation.

See: newGetStatusResponse smart constructor.

Constructors

GetStatusResponse'

Fields

trackingNumber :: Maybe Text

currentManifest :: Maybe Text

errorCount :: Maybe Int

creationDate :: Maybe ISO8601

logBucket :: Maybe Text

jobType :: Maybe JobType

signature :: Maybe Text

carrier :: Maybe Text

artifactList :: Maybe [Artifact]

progressMessage :: Maybe Text

locationMessage :: Maybe Text

logKey :: Maybe Text

signatureFileContents :: Maybe Text

locationCode :: Maybe Text

progressCode :: Maybe Text

jobId :: Maybe Text

httpStatus :: Int

The response's http status code.

newGetStatusResponse

Arguments

:: Int
$sel:httpStatus:GetStatusResponse'

-> GetStatusResponse

Create a value of GetStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trackingNumber:GetStatusResponse', getStatusResponse_trackingNumber - Undocumented member.

$sel:currentManifest:GetStatusResponse', getStatusResponse_currentManifest - Undocumented member.

$sel:errorCount:GetStatusResponse', getStatusResponse_errorCount - Undocumented member.

$sel:creationDate:GetStatusResponse', getStatusResponse_creationDate - Undocumented member.

$sel:logBucket:GetStatusResponse', getStatusResponse_logBucket - Undocumented member.

$sel:jobType:GetStatusResponse', getStatusResponse_jobType - Undocumented member.

$sel:signature:GetStatusResponse', getStatusResponse_signature - Undocumented member.

$sel:carrier:GetStatusResponse', getStatusResponse_carrier - Undocumented member.

$sel:artifactList:GetStatusResponse', getStatusResponse_artifactList - Undocumented member.

$sel:progressMessage:GetStatusResponse', getStatusResponse_progressMessage - Undocumented member.

$sel:locationMessage:GetStatusResponse', getStatusResponse_locationMessage - Undocumented member.

$sel:logKey:GetStatusResponse', getStatusResponse_logKey - Undocumented member.

$sel:signatureFileContents:GetStatusResponse', getStatusResponse_signatureFileContents - Undocumented member.

$sel:locationCode:GetStatusResponse', getStatusResponse_locationCode - Undocumented member.

$sel:progressCode:GetStatusResponse', getStatusResponse_progressCode - Undocumented member.

$sel:jobId:GetStatus', getStatusResponse_jobId - Undocumented member.

$sel:httpStatus:GetStatusResponse', getStatusResponse_httpStatus - The response's http status code.

Response Lenses

getStatusResponse_trackingNumber :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

getStatusResponse_currentManifest :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

getStatusResponse_errorCount :: Lens' GetStatusResponse (Maybe Int)

Undocumented member.

getStatusResponse_creationDate :: Lens' GetStatusResponse (Maybe UTCTime)

Undocumented member.

getStatusResponse_logBucket :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

getStatusResponse_jobType :: Lens' GetStatusResponse (Maybe JobType)

Undocumented member.

getStatusResponse_signature :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

getStatusResponse_carrier :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

getStatusResponse_artifactList :: Lens' GetStatusResponse (Maybe [Artifact])

Undocumented member.

getStatusResponse_progressMessage :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

getStatusResponse_locationMessage :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

getStatusResponse_logKey :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

getStatusResponse_signatureFileContents :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

getStatusResponse_locationCode :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

getStatusResponse_progressCode :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

getStatusResponse_jobId :: Lens' GetStatusResponse (Maybe Text)

Undocumented member.

getStatusResponse_httpStatus :: Lens' GetStatusResponse Int

The response's http status code.