Network.AWS.ImportExport.Types

Description

Service Configuration

defaultService :: Service

API version 2010-06-01 of the Amazon Import/Export SDK configuration.

Errors

_InvalidJobIdException :: AsError a => Getting (First ServiceError) a ServiceError

The JOBID was missing, not found, or not associated with the AWS account.

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

One or more parameters had an invalid value.

_ExpiredJobIdException :: AsError a => Getting (First ServiceError) a ServiceError

Indicates that the specified job has expired out of the system.

_InvalidFileSystemException :: AsError a => Getting (First ServiceError) a ServiceError

File system specified in export manifest is invalid.

_InvalidAccessKeyIdException :: AsError a => Getting (First ServiceError) a ServiceError

The AWS Access Key ID specified in the request did not match the manifest's accessKeyId value. The manifest and the request authentication must use the same AWS Access Key ID.

_UnableToUpdateJobIdException :: AsError a => Getting (First ServiceError) a ServiceError

AWS Import/Export cannot update the job

_UnableToCancelJobIdException :: AsError a => Getting (First ServiceError) a ServiceError

AWS Import/Export cannot cancel the job

_MultipleRegionsException :: AsError a => Getting (First ServiceError) a ServiceError

Your manifest file contained buckets from multiple regions. A job is restricted to buckets from one region. Please correct and resubmit.

_InvalidVersionException :: AsError a => Getting (First ServiceError) a ServiceError

The client tool version is invalid.

_MalformedManifestException :: AsError a => Getting (First ServiceError) a ServiceError

Your manifest is not well-formed.

_MissingParameterException :: AsError a => Getting (First ServiceError) a ServiceError

One or more required parameters was missing from the request.

_CanceledJobIdException :: AsError a => Getting (First ServiceError) a ServiceError

The specified job ID has been canceled and is no longer valid.

_BucketPermissionException :: AsError a => Getting (First ServiceError) a ServiceError

The account specified does not have the appropriate bucket permissions.

_NoSuchBucketException :: AsError a => Getting (First ServiceError) a ServiceError

The specified bucket does not exist. Create the specified bucket or change the manifest's bucket, exportBucket, or logBucket field to a bucket that the account, as specified by the manifest's Access Key ID, has write permissions to.

_InvalidAddressException :: AsError a => Getting (First ServiceError) a ServiceError

The address specified in the manifest is invalid.

_MissingCustomsException :: AsError a => Getting (First ServiceError) a ServiceError

One or more required customs parameters was missing from the manifest.

_InvalidManifestFieldException :: AsError a => Getting (First ServiceError) a ServiceError

One or more manifest fields was invalid. Please correct and resubmit.

_InvalidCustomsException :: AsError a => Getting (First ServiceError) a ServiceError

One or more customs parameters was invalid. Please correct and resubmit.

_MissingManifestFieldException :: AsError a => Getting (First ServiceError) a ServiceError

One or more required fields were missing from the manifest file. Please correct and resubmit.

_CreateJobQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Each account can create only a certain number of jobs per day. If you need to create more than this, please contact awsimportexport@amazon.com to explain your particular use case.

JobType

newtype JobType

Specifies whether the job to initiate is an import or export job.

Constructors

JobType'

Fields

fromJobType :: Text

Bundled Patterns

pattern JobType_Export :: JobType

pattern JobType_Import :: JobType

Artifact

data Artifact

A discrete item that contains the description and URL of an artifact (such as a PDF).

See: newArtifact smart constructor.

Constructors

Artifact'

Fields

url :: Maybe Text

description :: Maybe Text

newArtifact :: Artifact

Create a value of Artifact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:url:Artifact', artifact_url - Undocumented member.

$sel:description:Artifact', artifact_description - Undocumented member.

artifact_url :: Lens' Artifact (Maybe Text)

Undocumented member.

artifact_description :: Lens' Artifact (Maybe Text)

Undocumented member.

Job

data Job

Representation of a job returned by the ListJobs operation.

See: newJob smart constructor.

Constructors

Job'

Fields

jobType :: JobType

jobId :: Text

isCanceled :: Bool

creationDate :: ISO8601

newJob

Arguments

:: JobType
$sel:jobType:Job'

-> Text
$sel:jobId:Job'

-> Bool
$sel:isCanceled:Job'

-> UTCTime
$sel:creationDate:Job'

-> Job

Create a value of Job with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobType:Job', job_jobType - Undocumented member.

$sel:jobId:Job', job_jobId - Undocumented member.

$sel:isCanceled:Job', job_isCanceled - Undocumented member.

$sel:creationDate:Job', job_creationDate - Undocumented member.

job_jobType :: Lens' Job JobType

Undocumented member.

job_jobId :: Lens' Job Text

Undocumented member.

job_isCanceled :: Lens' Job Bool

Undocumented member.

job_creationDate :: Lens' Job UTCTime

Undocumented member.