Network.AWS.ImportExport.UpdateJob

Description

You use this operation to change the parameters specified in the original manifest file by supplying a new manifest file. The manifest file attached to this request replaces the original manifest file. You can only use the operation after a CreateJob request but before the data transfer starts and you can only use it on jobs you own.

Creating a Request

data UpdateJob

Input structure for the UpateJob operation.

See: newUpdateJob smart constructor.

Constructors

UpdateJob'

Fields

aPIVersion :: Maybe Text

jobId :: Text

manifest :: Text

jobType :: JobType

validateOnly :: Bool

newUpdateJob

Arguments

:: Text
$sel:jobId:UpdateJob'

-> Text
$sel:manifest:UpdateJob'

-> JobType
$sel:jobType:UpdateJob'

-> Bool
$sel:validateOnly:UpdateJob'

-> UpdateJob

Create a value of UpdateJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aPIVersion:UpdateJob', updateJob_aPIVersion - Undocumented member.

$sel:jobId:UpdateJob', updateJob_jobId - Undocumented member.

$sel:manifest:UpdateJob', updateJob_manifest - Undocumented member.

$sel:jobType:UpdateJob', updateJob_jobType - Undocumented member.

$sel:validateOnly:UpdateJob', updateJob_validateOnly - Undocumented member.

Request Lenses

updateJob_aPIVersion :: Lens' UpdateJob (Maybe Text)

Undocumented member.

updateJob_jobId :: Lens' UpdateJob Text

Undocumented member.

updateJob_manifest :: Lens' UpdateJob Text

Undocumented member.

updateJob_jobType :: Lens' UpdateJob JobType

Undocumented member.

updateJob_validateOnly :: Lens' UpdateJob Bool

Undocumented member.

Destructuring the Response

data UpdateJobResponse

Output structure for the UpateJob operation.

See: newUpdateJobResponse smart constructor.

Constructors

UpdateJobResponse'

Fields

warningMessage :: Maybe Text

artifactList :: Maybe [Artifact]

success :: Maybe Bool

httpStatus :: Int

The response's http status code.

newUpdateJobResponse

Arguments

:: Int
$sel:httpStatus:UpdateJobResponse'

-> UpdateJobResponse

Create a value of UpdateJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:warningMessage:UpdateJobResponse', updateJobResponse_warningMessage - Undocumented member.

$sel:artifactList:UpdateJobResponse', updateJobResponse_artifactList - Undocumented member.

$sel:success:UpdateJobResponse', updateJobResponse_success - Undocumented member.

$sel:httpStatus:UpdateJobResponse', updateJobResponse_httpStatus - The response's http status code.

Response Lenses

updateJobResponse_warningMessage :: Lens' UpdateJobResponse (Maybe Text)

Undocumented member.

updateJobResponse_artifactList :: Lens' UpdateJobResponse (Maybe [Artifact])

Undocumented member.

updateJobResponse_success :: Lens' UpdateJobResponse (Maybe Bool)

Undocumented member.

updateJobResponse_httpStatus :: Lens' UpdateJobResponse Int

The response's http status code.