Network.AWS.ImportExport.V2010_06_01.CreateJob

Description

This operation initiates the process of scheduling an upload or download of your data. You include in the request a manifest that describes the data transfer specifics. The response to the request includes a job ID, which you can use in other operations, a signature that you use to identify your storage device, and the address where you should ship your storage device.

Documentation

createJob

Arguments

:: JobType
_cjiJobType

-> Text
_cjiManifest

-> Bool
_cjiValidateOnly

-> CreateJob

Minimum specification for a CreateJob request.

data CreateJob

Constructors

CreateJob

Fields

_cjiJobType :: JobType

Specifies whether the job to initiate is an import or export job.

_cjiManifest :: Text

The UTF-8 encoded text of the manifest file.

_cjiValidateOnly :: Bool

Validate the manifest and parameter values in the request but do not actually create a job.

_cjiManifestAddendum :: Maybe Text

For internal use only.

cjiValidateOnly :: Lens' CreateJob Bool

cjiManifestAddendum :: Lens' CreateJob (Maybe Text)

cjiManifest :: Lens' CreateJob Text

cjiJobType :: Lens' CreateJob JobType

data CreateJobResponse

Constructors

CreateJobResponse

Fields

_cjoAwsShippingAddress :: Maybe Text

Address you ship your storage device to.

_cjoJobId :: Maybe Text

A unique identifier which refers to a particular job.

_cjoJobType :: Maybe JobType

Specifies whether the job to initiate is an import or export job.

_cjoSignature :: Maybe Text

An encrypted code used to authenticate the request and response, for example, "DV+TpDfx1TdSE9ktyK9kbDTVI=". Only use this value is you want to create the signature file yourself. Generally you should use the SignatureFileContents value.

_cjoSignatureFileContents :: Maybe Text

The actual text of the SIGNATURE file to be written to disk.

_cjoWarningMessage :: Maybe Text

An optional message notifying you of non-fatal issues with the job, such as use of an incompatible Amazon S3 bucket name.

cjoWarningMessage :: Lens' CreateJobResponse (Maybe Text)

cjoSignatureFileContents :: Lens' CreateJobResponse (Maybe Text)

cjoSignature :: Lens' CreateJobResponse (Maybe Text)

cjoJobType :: Lens' CreateJobResponse (Maybe JobType)

cjoJobId :: Lens' CreateJobResponse (Maybe Text)

cjoAwsShippingAddress :: Lens' CreateJobResponse (Maybe Text)