Network.AWS.ImportExport.V2010_06_01.GetStatus

Description

This operation returns information about a job, including where the job is in the processing pipeline, the status of the results, and the signature value associated with the job. You can only return information about jobs you own.

Documentation

data GetStatus

Constructors

GetStatus

Fields

_gsiJobId :: Text

A unique identifier which refers to a particular job.

gsiJobId :: Iso' GetStatus Text

data GetStatusResponse

Constructors

GetStatusResponse

Fields

_gsoAwsShippingAddress :: Maybe Text

Address you ship your storage device to.

_gsoCarrier :: Maybe Text

Name of the shipping company. This value is included when the LocationCode is Returned.

_gsoCreationDate :: Maybe ISO8601

Timestamp of the CreateJob request in ISO8601 date format. For example "2010-03-28T20:27:35Z".

_gsoCurrentManifest :: Maybe Text

The last manifest submitted, which will be used to process the job.

_gsoErrorCount :: Maybe Integer

Number of errors. We return this value when the ProgressCode is Success or SuccessWithErrors.

_gsoJobId :: Maybe Text

A unique identifier which refers to a particular job.

_gsoJobType :: Maybe JobType

Specifies whether the job to initiate is an import or export job.

_gsoLocationCode :: Maybe Text

A token representing the location of the storage device, such as AtAWS.

_gsoLocationMessage :: Maybe Text

A more human readable form of the physical location of the storage device.

_gsoLogBucket :: Maybe Text

Amazon S3 bucket for user logs.

_gsoLogKey :: Maybe Text

The key where the user logs were stored.

_gsoProgressCode :: Maybe Text

A token representing the state of the job, such as Started.

_gsoProgressMessage :: Maybe Text

A more human readable form of the job status.

_gsoSignature :: Maybe Text

An encrypted code used to authenticate the request and response, for example, "DV+TpDfx1TdSE9ktyK9kbDTVI=". Only use this value is you want to create the signature file yourself. Generally you should use the SignatureFileContents value.

_gsoSignatureFileContents :: Maybe Text

An encrypted code used to authenticate the request and response, for example, "DV+TpDfx1TdSE9ktyK9kbDTVI=". Only use this value is you want to create the signature file yourself. Generally you should use the SignatureFileContents value.

_gsoTrackingNumber :: Maybe Text

The shipping tracking number assigned by AWS Import/Export to the storage device when it's returned to you. We return this value when the LocationCode is Returned.

gsoTrackingNumber :: Lens' GetStatusResponse (Maybe Text)

gsoSignatureFileContents :: Lens' GetStatusResponse (Maybe Text)

gsoSignature :: Lens' GetStatusResponse (Maybe Text)

gsoProgressMessage :: Lens' GetStatusResponse (Maybe Text)

gsoProgressCode :: Lens' GetStatusResponse (Maybe Text)

gsoLogKey :: Lens' GetStatusResponse (Maybe Text)

gsoLogBucket :: Lens' GetStatusResponse (Maybe Text)

gsoLocationMessage :: Lens' GetStatusResponse (Maybe Text)

gsoLocationCode :: Lens' GetStatusResponse (Maybe Text)

gsoJobType :: Lens' GetStatusResponse (Maybe JobType)

gsoJobId :: Lens' GetStatusResponse (Maybe Text)

gsoErrorCount :: Lens' GetStatusResponse (Maybe Integer)

gsoCurrentManifest :: Lens' GetStatusResponse (Maybe Text)

gsoCreationDate :: Lens' GetStatusResponse (Maybe ISO8601)

gsoCarrier :: Lens' GetStatusResponse (Maybe Text)

gsoAwsShippingAddress :: Lens' GetStatusResponse (Maybe Text)