Network.AWS.ImportExport.V2010_06_01.UpdateJob

Description

You use this operation to change the parameters specified in the original manifest file by supplying a new manifest file. The manifest file attached to this request replaces the original manifest file. You can only use the operation after a CreateJob request but before the data transfer starts and you can only use it on jobs you own.

Documentation

data UpdateJob

Constructors

UpdateJob

Fields

_ujiJobId :: Text

A unique identifier which refers to a particular job.

_ujiJobType :: JobType

Specifies whether the job to initiate is an import or export job.

_ujiManifest :: Text

The UTF-8 encoded text of the manifest file.

_ujiValidateOnly :: Bool

Validate the manifest and parameter values in the request but do not actually create a job.

ujiValidateOnly :: Lens' UpdateJob Bool

ujiManifest :: Lens' UpdateJob Text

ujiJobType :: Lens' UpdateJob JobType

ujiJobId :: Lens' UpdateJob Text

data UpdateJobResponse

Constructors

UpdateJobResponse

Fields

_ujoSuccess :: Maybe Bool

Specifies whether (true) or not (false) AWS Import/Export updated your job.

_ujoWarningMessage :: Maybe Text

An optional message notifying you of non-fatal issues with the job, such as use of an incompatible Amazon S3 bucket name.

ujoWarningMessage :: Lens' UpdateJobResponse (Maybe Text)

ujoSuccess :: Lens' UpdateJobResponse (Maybe Bool)