Network.AWS.Inspector.CreateAssessmentTarget

Description

Creates a new assessment target using the ARN of the resource group that is generated by CreateResourceGroup. If resourceGroupArn is not specified, all EC2 instances in the current AWS account and region are included in the assessment target. If the service-linked role isn’t already registered, this action also creates and registers a service-linked role to grant Amazon Inspector access to AWS Services needed to perform security assessments. You can create up to 50 assessment targets per AWS account. You can run up to 500 concurrent agents per AWS account. For more information, see Amazon Inspector Assessment Targets.

Creating a Request

data CreateAssessmentTarget

See: newCreateAssessmentTarget smart constructor.

Constructors

CreateAssessmentTarget'

Fields

resourceGroupArn :: Maybe Text

The ARN that specifies the resource group that is used to create the assessment target. If resourceGroupArn is not specified, all EC2 instances in the current AWS account and region are included in the assessment target.

assessmentTargetName :: Text

The user-defined name that identifies the assessment target that you want to create. The name must be unique within the AWS account.

newCreateAssessmentTarget

Arguments

:: Text
$sel:assessmentTargetName:CreateAssessmentTarget'

-> CreateAssessmentTarget

Create a value of CreateAssessmentTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceGroupArn:CreateAssessmentTarget', createAssessmentTarget_resourceGroupArn - The ARN that specifies the resource group that is used to create the assessment target. If resourceGroupArn is not specified, all EC2 instances in the current AWS account and region are included in the assessment target.

$sel:assessmentTargetName:CreateAssessmentTarget', createAssessmentTarget_assessmentTargetName - The user-defined name that identifies the assessment target that you want to create. The name must be unique within the AWS account.

Request Lenses

createAssessmentTarget_resourceGroupArn :: Lens' CreateAssessmentTarget (Maybe Text)

The ARN that specifies the resource group that is used to create the assessment target. If resourceGroupArn is not specified, all EC2 instances in the current AWS account and region are included in the assessment target.

createAssessmentTarget_assessmentTargetName :: Lens' CreateAssessmentTarget Text

The user-defined name that identifies the assessment target that you want to create. The name must be unique within the AWS account.

Destructuring the Response

data CreateAssessmentTargetResponse

See: newCreateAssessmentTargetResponse smart constructor.

Constructors

CreateAssessmentTargetResponse'

Fields

httpStatus :: Int

The response's http status code.

assessmentTargetArn :: Text

The ARN that specifies the assessment target that is created.

newCreateAssessmentTargetResponse

Arguments

:: Int
$sel:httpStatus:CreateAssessmentTargetResponse'

-> Text
$sel:assessmentTargetArn:CreateAssessmentTargetResponse'

-> CreateAssessmentTargetResponse

Create a value of CreateAssessmentTargetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateAssessmentTargetResponse', createAssessmentTargetResponse_httpStatus - The response's http status code.

$sel:assessmentTargetArn:CreateAssessmentTargetResponse', createAssessmentTargetResponse_assessmentTargetArn - The ARN that specifies the assessment target that is created.

Response Lenses

createAssessmentTargetResponse_httpStatus :: Lens' CreateAssessmentTargetResponse Int

The response's http status code.

createAssessmentTargetResponse_assessmentTargetArn :: Lens' CreateAssessmentTargetResponse Text

The ARN that specifies the assessment target that is created.