Network.AWS.Inspector.CreateAssessmentTemplate

Description

Creates an assessment template for the assessment target that is specified by the ARN of the assessment target. If the service-linked role isn’t already registered, this action also creates and registers a service-linked role to grant Amazon Inspector access to AWS Services needed to perform security assessments.

Creating a Request

data CreateAssessmentTemplate

See: newCreateAssessmentTemplate smart constructor.

Constructors

CreateAssessmentTemplate'

Fields

userAttributesForFindings :: Maybe [Attribute]

The user-defined attributes that are assigned to every finding that is generated by the assessment run that uses this assessment template. An attribute is a key and value pair (an Attribute object). Within an assessment template, each key must be unique.

assessmentTargetArn :: Text

The ARN that specifies the assessment target for which you want to create the assessment template.

assessmentTemplateName :: Text

The user-defined name that identifies the assessment template that you want to create. You can create several assessment templates for an assessment target. The names of the assessment templates that correspond to a particular assessment target must be unique.

durationInSeconds :: Natural

The duration of the assessment run in seconds.

rulesPackageArns :: [Text]

The ARNs that specify the rules packages that you want to attach to the assessment template.

newCreateAssessmentTemplate

Arguments

:: Text
$sel:assessmentTargetArn:CreateAssessmentTemplate'

-> Text
$sel:assessmentTemplateName:CreateAssessmentTemplate'

-> Natural
$sel:durationInSeconds:CreateAssessmentTemplate'

-> CreateAssessmentTemplate

Create a value of CreateAssessmentTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userAttributesForFindings:CreateAssessmentTemplate', createAssessmentTemplate_userAttributesForFindings - The user-defined attributes that are assigned to every finding that is generated by the assessment run that uses this assessment template. An attribute is a key and value pair (an Attribute object). Within an assessment template, each key must be unique.

$sel:assessmentTargetArn:CreateAssessmentTemplate', createAssessmentTemplate_assessmentTargetArn - The ARN that specifies the assessment target for which you want to create the assessment template.

$sel:assessmentTemplateName:CreateAssessmentTemplate', createAssessmentTemplate_assessmentTemplateName - The user-defined name that identifies the assessment template that you want to create. You can create several assessment templates for an assessment target. The names of the assessment templates that correspond to a particular assessment target must be unique.

$sel:durationInSeconds:CreateAssessmentTemplate', createAssessmentTemplate_durationInSeconds - The duration of the assessment run in seconds.

$sel:rulesPackageArns:CreateAssessmentTemplate', createAssessmentTemplate_rulesPackageArns - The ARNs that specify the rules packages that you want to attach to the assessment template.

Request Lenses

createAssessmentTemplate_userAttributesForFindings :: Lens' CreateAssessmentTemplate (Maybe [Attribute])

The user-defined attributes that are assigned to every finding that is generated by the assessment run that uses this assessment template. An attribute is a key and value pair (an Attribute object). Within an assessment template, each key must be unique.

createAssessmentTemplate_assessmentTargetArn :: Lens' CreateAssessmentTemplate Text

The ARN that specifies the assessment target for which you want to create the assessment template.

createAssessmentTemplate_assessmentTemplateName :: Lens' CreateAssessmentTemplate Text

The user-defined name that identifies the assessment template that you want to create. You can create several assessment templates for an assessment target. The names of the assessment templates that correspond to a particular assessment target must be unique.

createAssessmentTemplate_durationInSeconds :: Lens' CreateAssessmentTemplate Natural

The duration of the assessment run in seconds.

createAssessmentTemplate_rulesPackageArns :: Lens' CreateAssessmentTemplate [Text]

The ARNs that specify the rules packages that you want to attach to the assessment template.

Destructuring the Response

data CreateAssessmentTemplateResponse

See: newCreateAssessmentTemplateResponse smart constructor.

Constructors

CreateAssessmentTemplateResponse'

Fields

httpStatus :: Int

The response's http status code.

assessmentTemplateArn :: Text

The ARN that specifies the assessment template that is created.

newCreateAssessmentTemplateResponse

Arguments

:: Int
$sel:httpStatus:CreateAssessmentTemplateResponse'

-> Text
$sel:assessmentTemplateArn:CreateAssessmentTemplateResponse'

-> CreateAssessmentTemplateResponse

Create a value of CreateAssessmentTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateAssessmentTemplateResponse', createAssessmentTemplateResponse_httpStatus - The response's http status code.

$sel:assessmentTemplateArn:CreateAssessmentTemplateResponse', createAssessmentTemplateResponse_assessmentTemplateArn - The ARN that specifies the assessment template that is created.

Response Lenses

createAssessmentTemplateResponse_httpStatus :: Lens' CreateAssessmentTemplateResponse Int

The response's http status code.

createAssessmentTemplateResponse_assessmentTemplateArn :: Lens' CreateAssessmentTemplateResponse Text

The ARN that specifies the assessment template that is created.