Network.AWS.Inspector.CreateResourceGroup

Description

Creates a resource group using the specified set of tags (key and value pairs) that are used to select the EC2 instances to be included in an Amazon Inspector assessment target. The created resource group is then used to create an Amazon Inspector assessment target. For more information, see CreateAssessmentTarget.

Creating a Request

data CreateResourceGroup

See: newCreateResourceGroup smart constructor.

Constructors

CreateResourceGroup'

Fields

resourceGroupTags :: NonEmpty ResourceGroupTag

A collection of keys and an array of possible values, '[{"key":"key1","values":["Value1","Value2"]},{"key":"Key2","values":["Value3"]}]'.

For example,'[{"key":"Name","values":["TestEC2Instance"]}]'.

newCreateResourceGroup

Arguments

:: NonEmpty ResourceGroupTag
$sel:resourceGroupTags:CreateResourceGroup'

-> CreateResourceGroup

Create a value of CreateResourceGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceGroupTags:CreateResourceGroup', createResourceGroup_resourceGroupTags - A collection of keys and an array of possible values, '[{"key":"key1","values":["Value1","Value2"]},{"key":"Key2","values":["Value3"]}]'.

For example,'[{"key":"Name","values":["TestEC2Instance"]}]'.

Request Lenses

createResourceGroup_resourceGroupTags :: Lens' CreateResourceGroup (NonEmpty ResourceGroupTag)

A collection of keys and an array of possible values, '[{"key":"key1","values":["Value1","Value2"]},{"key":"Key2","values":["Value3"]}]'.

For example,'[{"key":"Name","values":["TestEC2Instance"]}]'.

Destructuring the Response

data CreateResourceGroupResponse

See: newCreateResourceGroupResponse smart constructor.

Constructors

CreateResourceGroupResponse'

Fields

httpStatus :: Int

The response's http status code.

resourceGroupArn :: Text

The ARN that specifies the resource group that is created.

newCreateResourceGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateResourceGroupResponse'

-> Text
$sel:resourceGroupArn:CreateResourceGroupResponse'

-> CreateResourceGroupResponse

Create a value of CreateResourceGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateResourceGroupResponse', createResourceGroupResponse_httpStatus - The response's http status code.

$sel:resourceGroupArn:CreateResourceGroupResponse', createResourceGroupResponse_resourceGroupArn - The ARN that specifies the resource group that is created.

Response Lenses

createResourceGroupResponse_httpStatus :: Lens' CreateResourceGroupResponse Int

The response's http status code.

createResourceGroupResponse_resourceGroupArn :: Lens' CreateResourceGroupResponse Text

The ARN that specifies the resource group that is created.