Network.AWS.Inspector.DescribeAssessmentTargets

Description

Describes the assessment targets that are specified by the ARNs of the assessment targets.

Creating a Request

data DescribeAssessmentTargets

See: newDescribeAssessmentTargets smart constructor.

Constructors

DescribeAssessmentTargets'

Fields

assessmentTargetArns :: NonEmpty Text

The ARNs that specifies the assessment targets that you want to describe.

newDescribeAssessmentTargets

Arguments

:: NonEmpty Text
$sel:assessmentTargetArns:DescribeAssessmentTargets'

-> DescribeAssessmentTargets

Create a value of DescribeAssessmentTargets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assessmentTargetArns:DescribeAssessmentTargets', describeAssessmentTargets_assessmentTargetArns - The ARNs that specifies the assessment targets that you want to describe.

Request Lenses

describeAssessmentTargets_assessmentTargetArns :: Lens' DescribeAssessmentTargets (NonEmpty Text)

The ARNs that specifies the assessment targets that you want to describe.

Destructuring the Response

data DescribeAssessmentTargetsResponse

See: newDescribeAssessmentTargetsResponse smart constructor.

Constructors

DescribeAssessmentTargetsResponse'

Fields

httpStatus :: Int

The response's http status code.

assessmentTargets :: [AssessmentTarget]

Information about the assessment targets.

failedItems :: HashMap Text FailedItemDetails

Assessment target details that cannot be described. An error code is provided for each failed item.

newDescribeAssessmentTargetsResponse

Arguments

:: Int
$sel:httpStatus:DescribeAssessmentTargetsResponse'

-> DescribeAssessmentTargetsResponse

Create a value of DescribeAssessmentTargetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeAssessmentTargetsResponse', describeAssessmentTargetsResponse_httpStatus - The response's http status code.

$sel:assessmentTargets:DescribeAssessmentTargetsResponse', describeAssessmentTargetsResponse_assessmentTargets - Information about the assessment targets.

$sel:failedItems:DescribeAssessmentTargetsResponse', describeAssessmentTargetsResponse_failedItems - Assessment target details that cannot be described. An error code is provided for each failed item.

Response Lenses

describeAssessmentTargetsResponse_httpStatus :: Lens' DescribeAssessmentTargetsResponse Int

The response's http status code.

describeAssessmentTargetsResponse_assessmentTargets :: Lens' DescribeAssessmentTargetsResponse [AssessmentTarget]

Information about the assessment targets.

describeAssessmentTargetsResponse_failedItems :: Lens' DescribeAssessmentTargetsResponse (HashMap Text FailedItemDetails)

Assessment target details that cannot be described. An error code is provided for each failed item.