Network.AWS.Inspector.GetAssessmentReport

Description

Produces an assessment report that includes detailed and comprehensive results of a specified assessment run.

Creating a Request

data GetAssessmentReport

See: newGetAssessmentReport smart constructor.

Constructors

GetAssessmentReport'

Fields

assessmentRunArn :: Text

The ARN that specifies the assessment run for which you want to generate a report.

reportFileFormat :: ReportFileFormat

Specifies the file format (html or pdf) of the assessment report that you want to generate.

reportType :: ReportType

Specifies the type of the assessment report that you want to generate. There are two types of assessment reports: a finding report and a full report. For more information, see Assessment Reports.

newGetAssessmentReport

Arguments

:: Text
$sel:assessmentRunArn:GetAssessmentReport'

-> ReportFileFormat
$sel:reportFileFormat:GetAssessmentReport'

-> ReportType
$sel:reportType:GetAssessmentReport'

-> GetAssessmentReport

Create a value of GetAssessmentReport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assessmentRunArn:GetAssessmentReport', getAssessmentReport_assessmentRunArn - The ARN that specifies the assessment run for which you want to generate a report.

$sel:reportFileFormat:GetAssessmentReport', getAssessmentReport_reportFileFormat - Specifies the file format (html or pdf) of the assessment report that you want to generate.

$sel:reportType:GetAssessmentReport', getAssessmentReport_reportType - Specifies the type of the assessment report that you want to generate. There are two types of assessment reports: a finding report and a full report. For more information, see Assessment Reports.

Request Lenses

getAssessmentReport_assessmentRunArn :: Lens' GetAssessmentReport Text

The ARN that specifies the assessment run for which you want to generate a report.

getAssessmentReport_reportFileFormat :: Lens' GetAssessmentReport ReportFileFormat

Specifies the file format (html or pdf) of the assessment report that you want to generate.

getAssessmentReport_reportType :: Lens' GetAssessmentReport ReportType

Specifies the type of the assessment report that you want to generate. There are two types of assessment reports: a finding report and a full report. For more information, see Assessment Reports.

Destructuring the Response

data GetAssessmentReportResponse

See: newGetAssessmentReportResponse smart constructor.

Constructors

GetAssessmentReportResponse'

Fields

url :: Maybe Text

Specifies the URL where you can find the generated assessment report. This parameter is only returned if the report is successfully generated.

httpStatus :: Int

The response's http status code.

status :: ReportStatus

Specifies the status of the request to generate an assessment report.

newGetAssessmentReportResponse

Arguments

:: Int
$sel:httpStatus:GetAssessmentReportResponse'

-> ReportStatus
$sel:status:GetAssessmentReportResponse'

-> GetAssessmentReportResponse

Create a value of GetAssessmentReportResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:url:GetAssessmentReportResponse', getAssessmentReportResponse_url - Specifies the URL where you can find the generated assessment report. This parameter is only returned if the report is successfully generated.

$sel:httpStatus:GetAssessmentReportResponse', getAssessmentReportResponse_httpStatus - The response's http status code.

$sel:status:GetAssessmentReportResponse', getAssessmentReportResponse_status - Specifies the status of the request to generate an assessment report.

Response Lenses

getAssessmentReportResponse_url :: Lens' GetAssessmentReportResponse (Maybe Text)

Specifies the URL where you can find the generated assessment report. This parameter is only returned if the report is successfully generated.

getAssessmentReportResponse_httpStatus :: Lens' GetAssessmentReportResponse Int

The response's http status code.

getAssessmentReportResponse_status :: Lens' GetAssessmentReportResponse ReportStatus

Specifies the status of the request to generate an assessment report.