Network.AWS.Inspector.GetExclusionsPreview

Description

Retrieves the exclusions preview (a list of ExclusionPreview objects) specified by the preview token. You can obtain the preview token by running the CreateExclusionsPreview API.

Creating a Request

data GetExclusionsPreview

See: newGetExclusionsPreview smart constructor.

Constructors

GetExclusionsPreview'

Fields

nextToken :: Maybe Text

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the GetExclusionsPreviewRequest action. Subsequent calls to the action fill nextToken in the request with the value of nextToken from the previous response to continue listing data.

maxResults :: Maybe Int

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 100. The maximum value is 500.

locale :: Maybe Locale

The locale into which you want to translate the exclusion's title, description, and recommendation.

assessmentTemplateArn :: Text

The ARN that specifies the assessment template for which the exclusions preview was requested.

previewToken :: Text

The unique identifier associated of the exclusions preview.

newGetExclusionsPreview

Arguments

:: Text
$sel:assessmentTemplateArn:GetExclusionsPreview'

-> Text
$sel:previewToken:GetExclusionsPreview'

-> GetExclusionsPreview

Create a value of GetExclusionsPreview with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetExclusionsPreview', getExclusionsPreview_nextToken - You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the GetExclusionsPreviewRequest action. Subsequent calls to the action fill nextToken in the request with the value of nextToken from the previous response to continue listing data.

$sel:maxResults:GetExclusionsPreview', getExclusionsPreview_maxResults - You can use this parameter to indicate the maximum number of items you want in the response. The default value is 100. The maximum value is 500.

$sel:locale:GetExclusionsPreview', getExclusionsPreview_locale - The locale into which you want to translate the exclusion's title, description, and recommendation.

$sel:assessmentTemplateArn:GetExclusionsPreview', getExclusionsPreview_assessmentTemplateArn - The ARN that specifies the assessment template for which the exclusions preview was requested.

$sel:previewToken:GetExclusionsPreview', getExclusionsPreview_previewToken - The unique identifier associated of the exclusions preview.

Request Lenses

getExclusionsPreview_nextToken :: Lens' GetExclusionsPreview (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the GetExclusionsPreviewRequest action. Subsequent calls to the action fill nextToken in the request with the value of nextToken from the previous response to continue listing data.

getExclusionsPreview_maxResults :: Lens' GetExclusionsPreview (Maybe Int)

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 100. The maximum value is 500.

getExclusionsPreview_locale :: Lens' GetExclusionsPreview (Maybe Locale)

The locale into which you want to translate the exclusion's title, description, and recommendation.

getExclusionsPreview_assessmentTemplateArn :: Lens' GetExclusionsPreview Text

The ARN that specifies the assessment template for which the exclusions preview was requested.

getExclusionsPreview_previewToken :: Lens' GetExclusionsPreview Text

The unique identifier associated of the exclusions preview.

Destructuring the Response

data GetExclusionsPreviewResponse

See: newGetExclusionsPreviewResponse smart constructor.

Constructors

GetExclusionsPreviewResponse'

Fields

nextToken :: Maybe Text

When a response is generated, if there is more data to be listed, this parameters is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

exclusionPreviews :: Maybe [ExclusionPreview]

Information about the exclusions included in the preview.

httpStatus :: Int

The response's http status code.

previewStatus :: PreviewStatus

Specifies the status of the request to generate an exclusions preview.

newGetExclusionsPreviewResponse

Arguments

:: Int
$sel:httpStatus:GetExclusionsPreviewResponse'

-> PreviewStatus
$sel:previewStatus:GetExclusionsPreviewResponse'

-> GetExclusionsPreviewResponse

Create a value of GetExclusionsPreviewResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetExclusionsPreview', getExclusionsPreviewResponse_nextToken - When a response is generated, if there is more data to be listed, this parameters is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

$sel:exclusionPreviews:GetExclusionsPreviewResponse', getExclusionsPreviewResponse_exclusionPreviews - Information about the exclusions included in the preview.

$sel:httpStatus:GetExclusionsPreviewResponse', getExclusionsPreviewResponse_httpStatus - The response's http status code.

$sel:previewStatus:GetExclusionsPreviewResponse', getExclusionsPreviewResponse_previewStatus - Specifies the status of the request to generate an exclusions preview.

Response Lenses

getExclusionsPreviewResponse_nextToken :: Lens' GetExclusionsPreviewResponse (Maybe Text)

When a response is generated, if there is more data to be listed, this parameters is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

getExclusionsPreviewResponse_exclusionPreviews :: Lens' GetExclusionsPreviewResponse (Maybe [ExclusionPreview])

Information about the exclusions included in the preview.

getExclusionsPreviewResponse_httpStatus :: Lens' GetExclusionsPreviewResponse Int

The response's http status code.

getExclusionsPreviewResponse_previewStatus :: Lens' GetExclusionsPreviewResponse PreviewStatus

Specifies the status of the request to generate an exclusions preview.