Network.AWS.Inspector.Lens

Description

Operations

DeleteAssessmentTemplate

deleteAssessmentTemplate_assessmentTemplateArn :: Lens' DeleteAssessmentTemplate Text

The ARN that specifies the assessment template that you want to delete.

SubscribeToEvent

subscribeToEvent_resourceArn :: Lens' SubscribeToEvent Text

The ARN of the assessment template that is used during the event for which you want to receive SNS notifications.

subscribeToEvent_event :: Lens' SubscribeToEvent InspectorEvent

The event for which you want to receive SNS notifications.

subscribeToEvent_topicArn :: Lens' SubscribeToEvent Text

The ARN of the SNS topic to which the SNS notifications are sent.

StopAssessmentRun

stopAssessmentRun_stopAction :: Lens' StopAssessmentRun (Maybe StopAction)

An input option that can be set to either START_EVALUATION or SKIP_EVALUATION. START_EVALUATION (the default value), stops the AWS agent from collecting data and begins the results evaluation and the findings generation process. SKIP_EVALUATION cancels the assessment run immediately, after which no findings are generated.

stopAssessmentRun_assessmentRunArn :: Lens' StopAssessmentRun Text

The ARN of the assessment run that you want to stop.

StartAssessmentRun

startAssessmentRun_assessmentRunName :: Lens' StartAssessmentRun (Maybe Text)

You can specify the name for the assessment run. The name must be unique for the assessment template whose ARN is used to start the assessment run.

startAssessmentRun_assessmentTemplateArn :: Lens' StartAssessmentRun Text

The ARN of the assessment template of the assessment run that you want to start.

startAssessmentRunResponse_httpStatus :: Lens' StartAssessmentRunResponse Int

The response's http status code.

startAssessmentRunResponse_assessmentRunArn :: Lens' StartAssessmentRunResponse Text

The ARN of the assessment run that has been started.

ListFindings

listFindings_nextToken :: Lens' ListFindings (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListFindings action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

listFindings_maxResults :: Lens' ListFindings (Maybe Int)

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

listFindings_assessmentRunArns :: Lens' ListFindings (Maybe [Text])

The ARNs of the assessment runs that generate the findings that you want to list.

listFindings_filter :: Lens' ListFindings (Maybe FindingFilter)

You can use this parameter to specify a subset of data to be included in the action's response.

For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

listFindingsResponse_nextToken :: Lens' ListFindingsResponse (Maybe Text)

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

listFindingsResponse_httpStatus :: Lens' ListFindingsResponse Int

The response's http status code.

listFindingsResponse_findingArns :: Lens' ListFindingsResponse [Text]

A list of ARNs that specifies the findings returned by the action.

GetTelemetryMetadata

getTelemetryMetadata_assessmentRunArn :: Lens' GetTelemetryMetadata Text

The ARN that specifies the assessment run that has the telemetry data that you want to obtain.

getTelemetryMetadataResponse_httpStatus :: Lens' GetTelemetryMetadataResponse Int

The response's http status code.

getTelemetryMetadataResponse_telemetryMetadata :: Lens' GetTelemetryMetadataResponse [TelemetryMetadata]

Telemetry details.

DescribeAssessmentTargets

describeAssessmentTargets_assessmentTargetArns :: Lens' DescribeAssessmentTargets (NonEmpty Text)

The ARNs that specifies the assessment targets that you want to describe.

describeAssessmentTargetsResponse_httpStatus :: Lens' DescribeAssessmentTargetsResponse Int

The response's http status code.

describeAssessmentTargetsResponse_assessmentTargets :: Lens' DescribeAssessmentTargetsResponse [AssessmentTarget]

Information about the assessment targets.

describeAssessmentTargetsResponse_failedItems :: Lens' DescribeAssessmentTargetsResponse (HashMap Text FailedItemDetails)

Assessment target details that cannot be described. An error code is provided for each failed item.

ListAssessmentRuns

listAssessmentRuns_nextToken :: Lens' ListAssessmentRuns (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListAssessmentRuns action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

listAssessmentRuns_maxResults :: Lens' ListAssessmentRuns (Maybe Int)

You can use this parameter to indicate the maximum number of items that you want in the response. The default value is 10. The maximum value is 500.

listAssessmentRuns_filter :: Lens' ListAssessmentRuns (Maybe AssessmentRunFilter)

You can use this parameter to specify a subset of data to be included in the action's response.

For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

listAssessmentRuns_assessmentTemplateArns :: Lens' ListAssessmentRuns (Maybe [Text])

The ARNs that specify the assessment templates whose assessment runs you want to list.

listAssessmentRunsResponse_nextToken :: Lens' ListAssessmentRunsResponse (Maybe Text)

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

listAssessmentRunsResponse_httpStatus :: Lens' ListAssessmentRunsResponse Int

The response's http status code.

listAssessmentRunsResponse_assessmentRunArns :: Lens' ListAssessmentRunsResponse [Text]

A list of ARNs that specifies the assessment runs that are returned by the action.

DeleteAssessmentTarget

deleteAssessmentTarget_assessmentTargetArn :: Lens' DeleteAssessmentTarget Text

The ARN that specifies the assessment target that you want to delete.

ListAssessmentTargets

listAssessmentTargets_nextToken :: Lens' ListAssessmentTargets (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListAssessmentTargets action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

listAssessmentTargets_maxResults :: Lens' ListAssessmentTargets (Maybe Int)

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

listAssessmentTargets_filter :: Lens' ListAssessmentTargets (Maybe AssessmentTargetFilter)

You can use this parameter to specify a subset of data to be included in the action's response.

For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

listAssessmentTargetsResponse_nextToken :: Lens' ListAssessmentTargetsResponse (Maybe Text)

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

listAssessmentTargetsResponse_httpStatus :: Lens' ListAssessmentTargetsResponse Int

The response's http status code.

listAssessmentTargetsResponse_assessmentTargetArns :: Lens' ListAssessmentTargetsResponse [Text]

A list of ARNs that specifies the assessment targets that are returned by the action.

UpdateAssessmentTarget

updateAssessmentTarget_resourceGroupArn :: Lens' UpdateAssessmentTarget (Maybe Text)

The ARN of the resource group that is used to specify the new resource group to associate with the assessment target.

updateAssessmentTarget_assessmentTargetArn :: Lens' UpdateAssessmentTarget Text

The ARN of the assessment target that you want to update.

updateAssessmentTarget_assessmentTargetName :: Lens' UpdateAssessmentTarget Text

The name of the assessment target that you want to update.

DescribeResourceGroups

describeResourceGroups_resourceGroupArns :: Lens' DescribeResourceGroups (NonEmpty Text)

The ARN that specifies the resource group that you want to describe.

describeResourceGroupsResponse_httpStatus :: Lens' DescribeResourceGroupsResponse Int

The response's http status code.

describeResourceGroupsResponse_resourceGroups :: Lens' DescribeResourceGroupsResponse [ResourceGroup]

Information about a resource group.

describeResourceGroupsResponse_failedItems :: Lens' DescribeResourceGroupsResponse (HashMap Text FailedItemDetails)

Resource group details that cannot be described. An error code is provided for each failed item.

GetExclusionsPreview

getExclusionsPreview_nextToken :: Lens' GetExclusionsPreview (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the GetExclusionsPreviewRequest action. Subsequent calls to the action fill nextToken in the request with the value of nextToken from the previous response to continue listing data.

getExclusionsPreview_maxResults :: Lens' GetExclusionsPreview (Maybe Int)

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 100. The maximum value is 500.

getExclusionsPreview_locale :: Lens' GetExclusionsPreview (Maybe Locale)

The locale into which you want to translate the exclusion's title, description, and recommendation.

getExclusionsPreview_assessmentTemplateArn :: Lens' GetExclusionsPreview Text

The ARN that specifies the assessment template for which the exclusions preview was requested.

getExclusionsPreview_previewToken :: Lens' GetExclusionsPreview Text

The unique identifier associated of the exclusions preview.

getExclusionsPreviewResponse_nextToken :: Lens' GetExclusionsPreviewResponse (Maybe Text)

When a response is generated, if there is more data to be listed, this parameters is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

getExclusionsPreviewResponse_exclusionPreviews :: Lens' GetExclusionsPreviewResponse (Maybe [ExclusionPreview])

Information about the exclusions included in the preview.

getExclusionsPreviewResponse_httpStatus :: Lens' GetExclusionsPreviewResponse Int

The response's http status code.

getExclusionsPreviewResponse_previewStatus :: Lens' GetExclusionsPreviewResponse PreviewStatus

Specifies the status of the request to generate an exclusions preview.

AddAttributesToFindings

addAttributesToFindings_findingArns :: Lens' AddAttributesToFindings (NonEmpty Text)

The ARNs that specify the findings that you want to assign attributes to.

addAttributesToFindings_attributes :: Lens' AddAttributesToFindings [Attribute]

The array of attributes that you want to assign to specified findings.

addAttributesToFindingsResponse_httpStatus :: Lens' AddAttributesToFindingsResponse Int

The response's http status code.

addAttributesToFindingsResponse_failedItems :: Lens' AddAttributesToFindingsResponse (HashMap Text FailedItemDetails)

Attribute details that cannot be described. An error code is provided for each failed item.

CreateAssessmentTarget

createAssessmentTarget_resourceGroupArn :: Lens' CreateAssessmentTarget (Maybe Text)

The ARN that specifies the resource group that is used to create the assessment target. If resourceGroupArn is not specified, all EC2 instances in the current AWS account and region are included in the assessment target.

createAssessmentTarget_assessmentTargetName :: Lens' CreateAssessmentTarget Text

The user-defined name that identifies the assessment target that you want to create. The name must be unique within the AWS account.

createAssessmentTargetResponse_httpStatus :: Lens' CreateAssessmentTargetResponse Int

The response's http status code.

createAssessmentTargetResponse_assessmentTargetArn :: Lens' CreateAssessmentTargetResponse Text

The ARN that specifies the assessment target that is created.

PreviewAgents

previewAgents_nextToken :: Lens' PreviewAgents (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the PreviewAgents action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

previewAgents_maxResults :: Lens' PreviewAgents (Maybe Int)

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

previewAgents_previewAgentsArn :: Lens' PreviewAgents Text

The ARN of the assessment target whose agents you want to preview.

previewAgentsResponse_nextToken :: Lens' PreviewAgentsResponse (Maybe Text)

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

previewAgentsResponse_httpStatus :: Lens' PreviewAgentsResponse Int

The response's http status code.

previewAgentsResponse_agentPreviews :: Lens' PreviewAgentsResponse [AgentPreview]

The resulting list of agents.

ListExclusions

listExclusions_nextToken :: Lens' ListExclusions (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListExclusionsRequest action. Subsequent calls to the action fill nextToken in the request with the value of nextToken from the previous response to continue listing data.

listExclusions_maxResults :: Lens' ListExclusions (Maybe Int)

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 100. The maximum value is 500.

listExclusions_assessmentRunArn :: Lens' ListExclusions Text

The ARN of the assessment run that generated the exclusions that you want to list.

listExclusionsResponse_nextToken :: Lens' ListExclusionsResponse (Maybe Text)

When a response is generated, if there is more data to be listed, this parameters is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

listExclusionsResponse_httpStatus :: Lens' ListExclusionsResponse Int

The response's http status code.

listExclusionsResponse_exclusionArns :: Lens' ListExclusionsResponse [Text]

A list of exclusions' ARNs returned by the action.

CreateAssessmentTemplate

createAssessmentTemplate_userAttributesForFindings :: Lens' CreateAssessmentTemplate (Maybe [Attribute])

The user-defined attributes that are assigned to every finding that is generated by the assessment run that uses this assessment template. An attribute is a key and value pair (an Attribute object). Within an assessment template, each key must be unique.

createAssessmentTemplate_assessmentTargetArn :: Lens' CreateAssessmentTemplate Text

The ARN that specifies the assessment target for which you want to create the assessment template.

createAssessmentTemplate_assessmentTemplateName :: Lens' CreateAssessmentTemplate Text

The user-defined name that identifies the assessment template that you want to create. You can create several assessment templates for an assessment target. The names of the assessment templates that correspond to a particular assessment target must be unique.

createAssessmentTemplate_durationInSeconds :: Lens' CreateAssessmentTemplate Natural

The duration of the assessment run in seconds.

createAssessmentTemplate_rulesPackageArns :: Lens' CreateAssessmentTemplate [Text]

The ARNs that specify the rules packages that you want to attach to the assessment template.

createAssessmentTemplateResponse_httpStatus :: Lens' CreateAssessmentTemplateResponse Int

The response's http status code.

createAssessmentTemplateResponse_assessmentTemplateArn :: Lens' CreateAssessmentTemplateResponse Text

The ARN that specifies the assessment template that is created.

SetTagsForResource

setTagsForResource_tags :: Lens' SetTagsForResource (Maybe [Tag])

A collection of key and value pairs that you want to set to the assessment template.

setTagsForResource_resourceArn :: Lens' SetTagsForResource Text

The ARN of the assessment template that you want to set tags to.

DescribeCrossAccountAccessRole

describeCrossAccountAccessRoleResponse_httpStatus :: Lens' DescribeCrossAccountAccessRoleResponse Int

The response's http status code.

describeCrossAccountAccessRoleResponse_roleArn :: Lens' DescribeCrossAccountAccessRoleResponse Text

The ARN that specifies the IAM role that Amazon Inspector uses to access your AWS account.

describeCrossAccountAccessRoleResponse_valid :: Lens' DescribeCrossAccountAccessRoleResponse Bool

A Boolean value that specifies whether the IAM role has the necessary policies attached to enable Amazon Inspector to access your AWS account.

describeCrossAccountAccessRoleResponse_registeredAt :: Lens' DescribeCrossAccountAccessRoleResponse UTCTime

The date when the cross-account access role was registered.

DescribeExclusions

describeExclusions_locale :: Lens' DescribeExclusions (Maybe Locale)

The locale into which you want to translate the exclusion's title, description, and recommendation.

describeExclusions_exclusionArns :: Lens' DescribeExclusions (NonEmpty Text)

The list of ARNs that specify the exclusions that you want to describe.

describeExclusionsResponse_httpStatus :: Lens' DescribeExclusionsResponse Int

The response's http status code.

describeExclusionsResponse_exclusions :: Lens' DescribeExclusionsResponse (HashMap Text Exclusion)

Information about the exclusions.

describeExclusionsResponse_failedItems :: Lens' DescribeExclusionsResponse (HashMap Text FailedItemDetails)

Exclusion details that cannot be described. An error code is provided for each failed item.

ListAssessmentTemplates

listAssessmentTemplates_nextToken :: Lens' ListAssessmentTemplates (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListAssessmentTemplates action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

listAssessmentTemplates_maxResults :: Lens' ListAssessmentTemplates (Maybe Int)

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

listAssessmentTemplates_assessmentTargetArns :: Lens' ListAssessmentTemplates (Maybe [Text])

A list of ARNs that specifies the assessment targets whose assessment templates you want to list.

listAssessmentTemplates_filter :: Lens' ListAssessmentTemplates (Maybe AssessmentTemplateFilter)

You can use this parameter to specify a subset of data to be included in the action's response.

For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

listAssessmentTemplatesResponse_nextToken :: Lens' ListAssessmentTemplatesResponse (Maybe Text)

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

listAssessmentTemplatesResponse_httpStatus :: Lens' ListAssessmentTemplatesResponse Int

The response's http status code.

listAssessmentTemplatesResponse_assessmentTemplateArns :: Lens' ListAssessmentTemplatesResponse [Text]

A list of ARNs that specifies the assessment templates returned by the action.

ListAssessmentRunAgents

listAssessmentRunAgents_nextToken :: Lens' ListAssessmentRunAgents (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListAssessmentRunAgents action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

listAssessmentRunAgents_maxResults :: Lens' ListAssessmentRunAgents (Maybe Int)

You can use this parameter to indicate the maximum number of items that you want in the response. The default value is 10. The maximum value is 500.

listAssessmentRunAgents_filter :: Lens' ListAssessmentRunAgents (Maybe AgentFilter)

You can use this parameter to specify a subset of data to be included in the action's response.

For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

listAssessmentRunAgents_assessmentRunArn :: Lens' ListAssessmentRunAgents Text

The ARN that specifies the assessment run whose agents you want to list.

listAssessmentRunAgentsResponse_nextToken :: Lens' ListAssessmentRunAgentsResponse (Maybe Text)

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

listAssessmentRunAgentsResponse_httpStatus :: Lens' ListAssessmentRunAgentsResponse Int

The response's http status code.

listAssessmentRunAgentsResponse_assessmentRunAgents :: Lens' ListAssessmentRunAgentsResponse [AssessmentRunAgent]

A list of ARNs that specifies the agents returned by the action.

UnsubscribeFromEvent

unsubscribeFromEvent_resourceArn :: Lens' UnsubscribeFromEvent Text

The ARN of the assessment template that is used during the event for which you want to stop receiving SNS notifications.

unsubscribeFromEvent_event :: Lens' UnsubscribeFromEvent InspectorEvent

The event for which you want to stop receiving SNS notifications.

unsubscribeFromEvent_topicArn :: Lens' UnsubscribeFromEvent Text

The ARN of the SNS topic to which SNS notifications are sent.

DescribeAssessmentRuns

describeAssessmentRuns_assessmentRunArns :: Lens' DescribeAssessmentRuns (NonEmpty Text)

The ARN that specifies the assessment run that you want to describe.

describeAssessmentRunsResponse_httpStatus :: Lens' DescribeAssessmentRunsResponse Int

The response's http status code.

describeAssessmentRunsResponse_assessmentRuns :: Lens' DescribeAssessmentRunsResponse [AssessmentRun]

Information about the assessment run.

describeAssessmentRunsResponse_failedItems :: Lens' DescribeAssessmentRunsResponse (HashMap Text FailedItemDetails)

Assessment run details that cannot be described. An error code is provided for each failed item.

CreateExclusionsPreview

createExclusionsPreview_assessmentTemplateArn :: Lens' CreateExclusionsPreview Text

The ARN that specifies the assessment template for which you want to create an exclusions preview.

createExclusionsPreviewResponse_httpStatus :: Lens' CreateExclusionsPreviewResponse Int

The response's http status code.

createExclusionsPreviewResponse_previewToken :: Lens' CreateExclusionsPreviewResponse Text

Specifies the unique identifier of the requested exclusions preview. You can use the unique identifier to retrieve the exclusions preview when running the GetExclusionsPreview API.

DescribeRulesPackages

describeRulesPackages_locale :: Lens' DescribeRulesPackages (Maybe Locale)

The locale that you want to translate a rules package description into.

describeRulesPackages_rulesPackageArns :: Lens' DescribeRulesPackages (NonEmpty Text)

The ARN that specifies the rules package that you want to describe.

describeRulesPackagesResponse_httpStatus :: Lens' DescribeRulesPackagesResponse Int

The response's http status code.

describeRulesPackagesResponse_rulesPackages :: Lens' DescribeRulesPackagesResponse [RulesPackage]

Information about the rules package.

describeRulesPackagesResponse_failedItems :: Lens' DescribeRulesPackagesResponse (HashMap Text FailedItemDetails)

Rules package details that cannot be described. An error code is provided for each failed item.

RemoveAttributesFromFindings

removeAttributesFromFindings_findingArns :: Lens' RemoveAttributesFromFindings (NonEmpty Text)

The ARNs that specify the findings that you want to remove attributes from.

removeAttributesFromFindings_attributeKeys :: Lens' RemoveAttributesFromFindings [Text]

The array of attribute keys that you want to remove from specified findings.

removeAttributesFromFindingsResponse_httpStatus :: Lens' RemoveAttributesFromFindingsResponse Int

The response's http status code.

removeAttributesFromFindingsResponse_failedItems :: Lens' RemoveAttributesFromFindingsResponse (HashMap Text FailedItemDetails)

Attributes details that cannot be described. An error code is provided for each failed item.

CreateResourceGroup

createResourceGroup_resourceGroupTags :: Lens' CreateResourceGroup (NonEmpty ResourceGroupTag)

A collection of keys and an array of possible values, '[{"key":"key1","values":["Value1","Value2"]},{"key":"Key2","values":["Value3"]}]'.

For example,'[{"key":"Name","values":["TestEC2Instance"]}]'.

createResourceGroupResponse_httpStatus :: Lens' CreateResourceGroupResponse Int

The response's http status code.

createResourceGroupResponse_resourceGroupArn :: Lens' CreateResourceGroupResponse Text

The ARN that specifies the resource group that is created.

GetAssessmentReport

getAssessmentReport_assessmentRunArn :: Lens' GetAssessmentReport Text

The ARN that specifies the assessment run for which you want to generate a report.

getAssessmentReport_reportFileFormat :: Lens' GetAssessmentReport ReportFileFormat

Specifies the file format (html or pdf) of the assessment report that you want to generate.

getAssessmentReport_reportType :: Lens' GetAssessmentReport ReportType

Specifies the type of the assessment report that you want to generate. There are two types of assessment reports: a finding report and a full report. For more information, see Assessment Reports.

getAssessmentReportResponse_url :: Lens' GetAssessmentReportResponse (Maybe Text)

Specifies the URL where you can find the generated assessment report. This parameter is only returned if the report is successfully generated.

getAssessmentReportResponse_httpStatus :: Lens' GetAssessmentReportResponse Int

The response's http status code.

getAssessmentReportResponse_status :: Lens' GetAssessmentReportResponse ReportStatus

Specifies the status of the request to generate an assessment report.

DeleteAssessmentRun

deleteAssessmentRun_assessmentRunArn :: Lens' DeleteAssessmentRun Text

The ARN that specifies the assessment run that you want to delete.

ListEventSubscriptions

listEventSubscriptions_resourceArn :: Lens' ListEventSubscriptions (Maybe Text)

The ARN of the assessment template for which you want to list the existing event subscriptions.

listEventSubscriptions_nextToken :: Lens' ListEventSubscriptions (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListEventSubscriptions action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

listEventSubscriptions_maxResults :: Lens' ListEventSubscriptions (Maybe Int)

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

listEventSubscriptionsResponse_nextToken :: Lens' ListEventSubscriptionsResponse (Maybe Text)

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

listEventSubscriptionsResponse_httpStatus :: Lens' ListEventSubscriptionsResponse Int

The response's http status code.

listEventSubscriptionsResponse_subscriptions :: Lens' ListEventSubscriptionsResponse [Subscription]

Details of the returned event subscriptions.

ListRulesPackages

listRulesPackages_nextToken :: Lens' ListRulesPackages (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListRulesPackages action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

listRulesPackages_maxResults :: Lens' ListRulesPackages (Maybe Int)

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

listRulesPackagesResponse_nextToken :: Lens' ListRulesPackagesResponse (Maybe Text)

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

listRulesPackagesResponse_httpStatus :: Lens' ListRulesPackagesResponse Int

The response's http status code.

listRulesPackagesResponse_rulesPackageArns :: Lens' ListRulesPackagesResponse [Text]

The list of ARNs that specifies the rules packages returned by the action.

RegisterCrossAccountAccessRole

registerCrossAccountAccessRole_roleArn :: Lens' RegisterCrossAccountAccessRole Text

The ARN of the IAM role that grants Amazon Inspector access to AWS Services needed to perform security assessments.

DescribeFindings

describeFindings_locale :: Lens' DescribeFindings (Maybe Locale)

The locale into which you want to translate a finding description, recommendation, and the short description that identifies the finding.

describeFindings_findingArns :: Lens' DescribeFindings (NonEmpty Text)

The ARN that specifies the finding that you want to describe.

describeFindingsResponse_httpStatus :: Lens' DescribeFindingsResponse Int

The response's http status code.

describeFindingsResponse_findings :: Lens' DescribeFindingsResponse [Finding]

Information about the finding.

describeFindingsResponse_failedItems :: Lens' DescribeFindingsResponse (HashMap Text FailedItemDetails)

Finding details that cannot be described. An error code is provided for each failed item.

DescribeAssessmentTemplates

describeAssessmentTemplates_assessmentTemplateArns :: Lens' DescribeAssessmentTemplates (NonEmpty Text)

Undocumented member.

describeAssessmentTemplatesResponse_httpStatus :: Lens' DescribeAssessmentTemplatesResponse Int

The response's http status code.

describeAssessmentTemplatesResponse_assessmentTemplates :: Lens' DescribeAssessmentTemplatesResponse [AssessmentTemplate]

Information about the assessment templates.

describeAssessmentTemplatesResponse_failedItems :: Lens' DescribeAssessmentTemplatesResponse (HashMap Text FailedItemDetails)

Assessment template details that cannot be described. An error code is provided for each failed item.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The ARN that specifies the assessment template whose tags you want to list.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse [Tag]

A collection of key and value pairs.

Types

AgentFilter

agentFilter_agentHealths :: Lens' AgentFilter [AgentHealth]

The current health state of the agent. Values can be set to HEALTHY or UNHEALTHY.

agentFilter_agentHealthCodes :: Lens' AgentFilter [AgentHealthCode]

The detailed health state of the agent. Values can be set to IDLE, RUNNING, SHUTDOWN, UNHEALTHY, THROTTLED, and UNKNOWN.

AgentPreview

agentPreview_hostname :: Lens' AgentPreview (Maybe Text)

The hostname of the EC2 instance on which the Amazon Inspector Agent is installed.

agentPreview_agentVersion :: Lens' AgentPreview (Maybe Text)

The version of the Amazon Inspector Agent.

agentPreview_kernelVersion :: Lens' AgentPreview (Maybe Text)

The kernel version of the operating system running on the EC2 instance on which the Amazon Inspector Agent is installed.

agentPreview_operatingSystem :: Lens' AgentPreview (Maybe Text)

The operating system running on the EC2 instance on which the Amazon Inspector Agent is installed.

agentPreview_agentHealth :: Lens' AgentPreview (Maybe AgentHealth)

The health status of the Amazon Inspector Agent.

agentPreview_autoScalingGroup :: Lens' AgentPreview (Maybe Text)

The Auto Scaling group for the EC2 instance where the agent is installed.

agentPreview_ipv4Address :: Lens' AgentPreview (Maybe Text)

The IP address of the EC2 instance on which the Amazon Inspector Agent is installed.

agentPreview_agentId :: Lens' AgentPreview Text

The ID of the EC2 instance where the agent is installed.

AssessmentRun

assessmentRun_startedAt :: Lens' AssessmentRun (Maybe UTCTime)

The time when StartAssessmentRun was called.

assessmentRun_completedAt :: Lens' AssessmentRun (Maybe UTCTime)

The assessment run completion time that corresponds to the rules packages evaluation completion time or failure.

assessmentRun_arn :: Lens' AssessmentRun Text

The ARN of the assessment run.

assessmentRun_name :: Lens' AssessmentRun Text

The auto-generated name for the assessment run.

assessmentRun_assessmentTemplateArn :: Lens' AssessmentRun Text

The ARN of the assessment template that is associated with the assessment run.

assessmentRun_state :: Lens' AssessmentRun AssessmentRunState

The state of the assessment run.

assessmentRun_durationInSeconds :: Lens' AssessmentRun Natural

The duration of the assessment run.

assessmentRun_rulesPackageArns :: Lens' AssessmentRun (NonEmpty Text)

The rules packages selected for the assessment run.

assessmentRun_userAttributesForFindings :: Lens' AssessmentRun [Attribute]

The user-defined attributes that are assigned to every generated finding.

assessmentRun_createdAt :: Lens' AssessmentRun UTCTime

The time when StartAssessmentRun was called.

assessmentRun_stateChangedAt :: Lens' AssessmentRun UTCTime

The last time when the assessment run's state changed.

assessmentRun_dataCollected :: Lens' AssessmentRun Bool

A Boolean value (true or false) that specifies whether the process of collecting data from the agents is completed.

assessmentRun_stateChanges :: Lens' AssessmentRun [AssessmentRunStateChange]

A list of the assessment run state changes.

assessmentRun_notifications :: Lens' AssessmentRun [AssessmentRunNotification]

A list of notifications for the event subscriptions. A notification about a particular generated finding is added to this list only once.

assessmentRun_findingCounts :: Lens' AssessmentRun (HashMap Severity Int)

Provides a total count of generated findings per severity.

AssessmentRunAgent

assessmentRunAgent_agentHealthDetails :: Lens' AssessmentRunAgent (Maybe Text)

The description for the agent health code.

assessmentRunAgent_autoScalingGroup :: Lens' AssessmentRunAgent (Maybe Text)

The Auto Scaling group of the EC2 instance that is specified by the agent ID.

assessmentRunAgent_agentId :: Lens' AssessmentRunAgent Text

The AWS account of the EC2 instance where the agent is installed.

assessmentRunAgent_assessmentRunArn :: Lens' AssessmentRunAgent Text

The ARN of the assessment run that is associated with the agent.

assessmentRunAgent_agentHealth :: Lens' AssessmentRunAgent AgentHealth

The current health state of the agent.

assessmentRunAgent_agentHealthCode :: Lens' AssessmentRunAgent AgentHealthCode

The detailed health state of the agent.

assessmentRunAgent_telemetryMetadata :: Lens' AssessmentRunAgent [TelemetryMetadata]

The Amazon Inspector application data metrics that are collected by the agent.

AssessmentRunFilter

assessmentRunFilter_states :: Lens' AssessmentRunFilter (Maybe [AssessmentRunState])

For a record to match a filter, one of the values specified for this data type property must be the exact match of the value of the assessmentRunState property of the AssessmentRun data type.

assessmentRunFilter_durationRange :: Lens' AssessmentRunFilter (Maybe DurationRange)

For a record to match a filter, the value that is specified for this data type property must inclusively match any value between the specified minimum and maximum values of the durationInSeconds property of the AssessmentRun data type.

assessmentRunFilter_rulesPackageArns :: Lens' AssessmentRunFilter (Maybe [Text])

For a record to match a filter, the value that is specified for this data type property must be contained in the list of values of the rulesPackages property of the AssessmentRun data type.

assessmentRunFilter_stateChangeTimeRange :: Lens' AssessmentRunFilter (Maybe TimestampRange)

For a record to match a filter, the value that is specified for this data type property must match the stateChangedAt property of the AssessmentRun data type.

assessmentRunFilter_startTimeRange :: Lens' AssessmentRunFilter (Maybe TimestampRange)

For a record to match a filter, the value that is specified for this data type property must inclusively match any value between the specified minimum and maximum values of the startTime property of the AssessmentRun data type.

assessmentRunFilter_namePattern :: Lens' AssessmentRunFilter (Maybe Text)

For a record to match a filter, an explicit value or a string containing a wildcard that is specified for this data type property must match the value of the assessmentRunName property of the AssessmentRun data type.

assessmentRunFilter_completionTimeRange :: Lens' AssessmentRunFilter (Maybe TimestampRange)

For a record to match a filter, the value that is specified for this data type property must inclusively match any value between the specified minimum and maximum values of the completedAt property of the AssessmentRun data type.

AssessmentRunNotification

assessmentRunNotification_message :: Lens' AssessmentRunNotification (Maybe Text)

The message included in the notification.

assessmentRunNotification_snsPublishStatusCode :: Lens' AssessmentRunNotification (Maybe AssessmentRunNotificationSnsStatusCode)

The status code of the SNS notification.

assessmentRunNotification_snsTopicArn :: Lens' AssessmentRunNotification (Maybe Text)

The SNS topic to which the SNS notification is sent.

assessmentRunNotification_date :: Lens' AssessmentRunNotification UTCTime

The date of the notification.

assessmentRunNotification_event :: Lens' AssessmentRunNotification InspectorEvent

The event for which a notification is sent.

assessmentRunNotification_error :: Lens' AssessmentRunNotification Bool

The Boolean value that specifies whether the notification represents an error.

AssessmentRunStateChange

assessmentRunStateChange_stateChangedAt :: Lens' AssessmentRunStateChange UTCTime

The last time the assessment run state changed.

assessmentRunStateChange_state :: Lens' AssessmentRunStateChange AssessmentRunState

The assessment run state.

AssessmentTarget

assessmentTarget_resourceGroupArn :: Lens' AssessmentTarget (Maybe Text)

The ARN that specifies the resource group that is associated with the assessment target.

assessmentTarget_arn :: Lens' AssessmentTarget Text

The ARN that specifies the Amazon Inspector assessment target.

assessmentTarget_name :: Lens' AssessmentTarget Text

The name of the Amazon Inspector assessment target.

assessmentTarget_createdAt :: Lens' AssessmentTarget UTCTime

The time at which the assessment target is created.

assessmentTarget_updatedAt :: Lens' AssessmentTarget UTCTime

The time at which UpdateAssessmentTarget is called.

AssessmentTargetFilter

assessmentTargetFilter_assessmentTargetNamePattern :: Lens' AssessmentTargetFilter (Maybe Text)

For a record to match a filter, an explicit value or a string that contains a wildcard that is specified for this data type property must match the value of the assessmentTargetName property of the AssessmentTarget data type.

AssessmentTemplate

assessmentTemplate_lastAssessmentRunArn :: Lens' AssessmentTemplate (Maybe Text)

The Amazon Resource Name (ARN) of the most recent assessment run associated with this assessment template. This value exists only when the value of assessmentRunCount is greaterpa than zero.

assessmentTemplate_arn :: Lens' AssessmentTemplate Text

The ARN of the assessment template.

assessmentTemplate_name :: Lens' AssessmentTemplate Text

The name of the assessment template.

assessmentTemplate_assessmentTargetArn :: Lens' AssessmentTemplate Text

The ARN of the assessment target that corresponds to this assessment template.

assessmentTemplate_durationInSeconds :: Lens' AssessmentTemplate Natural

The duration in seconds specified for this assessment template. The default value is 3600 seconds (one hour). The maximum value is 86400 seconds (one day).

assessmentTemplate_rulesPackageArns :: Lens' AssessmentTemplate [Text]

The rules packages that are specified for this assessment template.

assessmentTemplate_userAttributesForFindings :: Lens' AssessmentTemplate [Attribute]

The user-defined attributes that are assigned to every generated finding from the assessment run that uses this assessment template.

assessmentTemplate_assessmentRunCount :: Lens' AssessmentTemplate Int

The number of existing assessment runs associated with this assessment template. This value can be zero or a positive integer.

assessmentTemplate_createdAt :: Lens' AssessmentTemplate UTCTime

The time at which the assessment template is created.

AssessmentTemplateFilter

assessmentTemplateFilter_durationRange :: Lens' AssessmentTemplateFilter (Maybe DurationRange)

For a record to match a filter, the value specified for this data type property must inclusively match any value between the specified minimum and maximum values of the durationInSeconds property of the AssessmentTemplate data type.

assessmentTemplateFilter_rulesPackageArns :: Lens' AssessmentTemplateFilter (Maybe [Text])

For a record to match a filter, the values that are specified for this data type property must be contained in the list of values of the rulesPackageArns property of the AssessmentTemplate data type.

assessmentTemplateFilter_namePattern :: Lens' AssessmentTemplateFilter (Maybe Text)

For a record to match a filter, an explicit value or a string that contains a wildcard that is specified for this data type property must match the value of the assessmentTemplateName property of the AssessmentTemplate data type.

AssetAttributes

assetAttributes_hostname :: Lens' AssetAttributes (Maybe Text)

The hostname of the EC2 instance where the finding is generated.

assetAttributes_agentId :: Lens' AssetAttributes (Maybe Text)

The ID of the agent that is installed on the EC2 instance where the finding is generated.

assetAttributes_amiId :: Lens' AssetAttributes (Maybe Text)

The ID of the Amazon Machine Image (AMI) that is installed on the EC2 instance where the finding is generated.

assetAttributes_tags :: Lens' AssetAttributes (Maybe [Tag])

The tags related to the EC2 instance where the finding is generated.

assetAttributes_ipv4Addresses :: Lens' AssetAttributes (Maybe [Text])

The list of IP v4 addresses of the EC2 instance where the finding is generated.

assetAttributes_networkInterfaces :: Lens' AssetAttributes (Maybe [NetworkInterface])

An array of the network interfaces interacting with the EC2 instance where the finding is generated.

assetAttributes_autoScalingGroup :: Lens' AssetAttributes (Maybe Text)

The Auto Scaling group of the EC2 instance where the finding is generated.

assetAttributes_schemaVersion :: Lens' AssetAttributes Natural

The schema version of this data type.

Attribute

attribute_value :: Lens' Attribute (Maybe Text)

The value assigned to the attribute key.

attribute_key :: Lens' Attribute Text

The attribute key.

DurationRange

durationRange_minSeconds :: Lens' DurationRange (Maybe Natural)

The minimum value of the duration range. Must be greater than zero.

durationRange_maxSeconds :: Lens' DurationRange (Maybe Natural)

The maximum value of the duration range. Must be less than or equal to 604800 seconds (1 week).

EventSubscription

eventSubscription_event :: Lens' EventSubscription InspectorEvent

The event for which Amazon Simple Notification Service (SNS) notifications are sent.

eventSubscription_subscribedAt :: Lens' EventSubscription UTCTime

The time at which SubscribeToEvent is called.

Exclusion

exclusion_attributes :: Lens' Exclusion (Maybe [Attribute])

The system-defined attributes for the exclusion.

exclusion_arn :: Lens' Exclusion Text

The ARN that specifies the exclusion.

exclusion_title :: Lens' Exclusion Text

The name of the exclusion.

exclusion_description :: Lens' Exclusion Text

The description of the exclusion.

exclusion_recommendation :: Lens' Exclusion Text

The recommendation for the exclusion.

exclusion_scopes :: Lens' Exclusion (NonEmpty Scope)

The AWS resources for which the exclusion pertains.

ExclusionPreview

exclusionPreview_attributes :: Lens' ExclusionPreview (Maybe [Attribute])

The system-defined attributes for the exclusion preview.

exclusionPreview_title :: Lens' ExclusionPreview Text

The name of the exclusion preview.

exclusionPreview_description :: Lens' ExclusionPreview Text

The description of the exclusion preview.

exclusionPreview_recommendation :: Lens' ExclusionPreview Text

The recommendation for the exclusion preview.

exclusionPreview_scopes :: Lens' ExclusionPreview (NonEmpty Scope)

The AWS resources for which the exclusion preview pertains.

FailedItemDetails

failedItemDetails_failureCode :: Lens' FailedItemDetails FailedItemErrorCode

The status code of a failed item.

failedItemDetails_retryable :: Lens' FailedItemDetails Bool

Indicates whether you can immediately retry a request for this item for a specified resource.

Finding

finding_assetAttributes :: Lens' Finding (Maybe AssetAttributes)

A collection of attributes of the host from which the finding is generated.

finding_severity :: Lens' Finding (Maybe Severity)

The finding severity. Values can be set to High, Medium, Low, and Informational.

finding_assetType :: Lens' Finding (Maybe AssetType)

The type of the host from which the finding is generated.

finding_numericSeverity :: Lens' Finding (Maybe Double)

The numeric value of the finding severity.

finding_title :: Lens' Finding (Maybe Text)

The name of the finding.

finding_id :: Lens' Finding (Maybe Text)

The ID of the finding.

finding_service :: Lens' Finding (Maybe Text)

The data element is set to "Inspector".

finding_serviceAttributes :: Lens' Finding (Maybe InspectorServiceAttributes)

This data type is used in the Finding data type.

finding_confidence :: Lens' Finding (Maybe Natural)

This data element is currently not used.

finding_recommendation :: Lens' Finding (Maybe Text)

The recommendation for the finding.

finding_indicatorOfCompromise :: Lens' Finding (Maybe Bool)

This data element is currently not used.

finding_description :: Lens' Finding (Maybe Text)

The description of the finding.

finding_schemaVersion :: Lens' Finding (Maybe Natural)

The schema version of this data type.

finding_arn :: Lens' Finding Text

The ARN that specifies the finding.

finding_attributes :: Lens' Finding [Attribute]

The system-defined attributes for the finding.

finding_userAttributes :: Lens' Finding [Attribute]

The user-defined attributes that are assigned to the finding.

finding_createdAt :: Lens' Finding UTCTime

The time when the finding was generated.

finding_updatedAt :: Lens' Finding UTCTime

The time when AddAttributesToFindings is called.

FindingFilter

findingFilter_agentIds :: Lens' FindingFilter (Maybe [Text])

For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the agentId property of the Finding data type.

findingFilter_rulesPackageArns :: Lens' FindingFilter (Maybe [Text])

For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the rulesPackageArn property of the Finding data type.

findingFilter_creationTimeRange :: Lens' FindingFilter (Maybe TimestampRange)

The time range during which the finding is generated.

findingFilter_severities :: Lens' FindingFilter (Maybe [Severity])

For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the severity property of the Finding data type.

findingFilter_attributes :: Lens' FindingFilter (Maybe [Attribute])

For a record to match a filter, the list of values that are specified for this data type property must be contained in the list of values of the attributes property of the Finding data type.

findingFilter_userAttributes :: Lens' FindingFilter (Maybe [Attribute])

For a record to match a filter, the value that is specified for this data type property must be contained in the list of values of the userAttributes property of the Finding data type.

findingFilter_autoScalingGroups :: Lens' FindingFilter (Maybe [Text])

For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the autoScalingGroup property of the Finding data type.

findingFilter_ruleNames :: Lens' FindingFilter (Maybe [Text])

For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the ruleName property of the Finding data type.

InspectorServiceAttributes

inspectorServiceAttributes_rulesPackageArn :: Lens' InspectorServiceAttributes (Maybe Text)

The ARN of the rules package that is used to generate the finding.

inspectorServiceAttributes_assessmentRunArn :: Lens' InspectorServiceAttributes (Maybe Text)

The ARN of the assessment run during which the finding is generated.

inspectorServiceAttributes_schemaVersion :: Lens' InspectorServiceAttributes Natural

The schema version of this data type.

NetworkInterface

networkInterface_privateIpAddresses :: Lens' NetworkInterface (Maybe [PrivateIp])

A list of the private IP addresses associated with the network interface. Includes the privateDnsName and privateIpAddress.

networkInterface_ipv6Addresses :: Lens' NetworkInterface (Maybe [Text])

The IP addresses associated with the network interface.

networkInterface_securityGroups :: Lens' NetworkInterface (Maybe [SecurityGroup])

A list of the security groups associated with the network interface. Includes the groupId and groupName.

networkInterface_publicDnsName :: Lens' NetworkInterface (Maybe Text)

The name of a public DNS associated with the network interface.

networkInterface_subnetId :: Lens' NetworkInterface (Maybe Text)

The ID of a subnet associated with the network interface.

networkInterface_networkInterfaceId :: Lens' NetworkInterface (Maybe Text)

The ID of the network interface.

networkInterface_privateDnsName :: Lens' NetworkInterface (Maybe Text)

The name of a private DNS associated with the network interface.

networkInterface_vpcId :: Lens' NetworkInterface (Maybe Text)

The ID of a VPC associated with the network interface.

networkInterface_publicIp :: Lens' NetworkInterface (Maybe Text)

The public IP address from which the network interface is reachable.

networkInterface_privateIpAddress :: Lens' NetworkInterface (Maybe Text)

The private IP address associated with the network interface.

PrivateIp

privateIp_privateDnsName :: Lens' PrivateIp (Maybe Text)

The DNS name of the private IP address.

privateIp_privateIpAddress :: Lens' PrivateIp (Maybe Text)

The full IP address of the network inteface.

ResourceGroup

resourceGroup_arn :: Lens' ResourceGroup Text

The ARN of the resource group.

resourceGroup_tags :: Lens' ResourceGroup (NonEmpty ResourceGroupTag)

The tags (key and value pairs) of the resource group. This data type property is used in the CreateResourceGroup action.

resourceGroup_createdAt :: Lens' ResourceGroup UTCTime

The time at which resource group is created.

ResourceGroupTag

resourceGroupTag_value :: Lens' ResourceGroupTag (Maybe Text)

The value assigned to a tag key.

resourceGroupTag_key :: Lens' ResourceGroupTag Text

A tag key.

RulesPackage

rulesPackage_description :: Lens' RulesPackage (Maybe Text)

The description of the rules package.

rulesPackage_arn :: Lens' RulesPackage Text

The ARN of the rules package.

rulesPackage_name :: Lens' RulesPackage Text

The name of the rules package.

rulesPackage_version :: Lens' RulesPackage Text

The version ID of the rules package.

rulesPackage_provider :: Lens' RulesPackage Text

The provider of the rules package.

Scope

scope_key :: Lens' Scope (Maybe ScopeType)

The type of the scope.

scope_value :: Lens' Scope (Maybe Text)

The resource identifier for the specified scope type.

SecurityGroup

securityGroup_groupName :: Lens' SecurityGroup (Maybe Text)

The name of the security group.

securityGroup_groupId :: Lens' SecurityGroup (Maybe Text)

The ID of the security group.

Subscription

subscription_resourceArn :: Lens' Subscription Text

The ARN of the assessment template that is used during the event for which the SNS notification is sent.

subscription_topicArn :: Lens' Subscription Text

The ARN of the Amazon Simple Notification Service (SNS) topic to which the SNS notifications are sent.

subscription_eventSubscriptions :: Lens' Subscription (NonEmpty EventSubscription)

The list of existing event subscriptions.

Tag

tag_value :: Lens' Tag (Maybe Text)

A value assigned to a tag key.

tag_key :: Lens' Tag Text

A tag key.

TelemetryMetadata

telemetryMetadata_dataSize :: Lens' TelemetryMetadata (Maybe Integer)

The data size of messages that the agent sends to the Amazon Inspector service.

telemetryMetadata_messageType :: Lens' TelemetryMetadata Text

A specific type of behavioral data that is collected by the agent.

telemetryMetadata_count :: Lens' TelemetryMetadata Integer

The count of messages that the agent sends to the Amazon Inspector service.

TimestampRange

timestampRange_beginDate :: Lens' TimestampRange (Maybe UTCTime)

The minimum value of the timestamp range.

timestampRange_endDate :: Lens' TimestampRange (Maybe UTCTime)

The maximum value of the timestamp range.