Network.AWS.Inspector.ListAssessmentRuns

Description

Lists the assessment runs that correspond to the assessment templates that are specified by the ARNs of the assessment templates.

This operation returns paginated results.

Creating a Request

data ListAssessmentRuns

See: newListAssessmentRuns smart constructor.

Constructors

ListAssessmentRuns'

Fields

nextToken :: Maybe Text

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListAssessmentRuns action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

filter' :: Maybe AssessmentRunFilter

You can use this parameter to specify a subset of data to be included in the action's response.

For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

assessmentTemplateArns :: Maybe [Text]

The ARNs that specify the assessment templates whose assessment runs you want to list.

maxResults :: Maybe Int

You can use this parameter to indicate the maximum number of items that you want in the response. The default value is 10. The maximum value is 500.

newListAssessmentRuns :: ListAssessmentRuns

Create a value of ListAssessmentRuns with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAssessmentRuns', listAssessmentRuns_nextToken - You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListAssessmentRuns action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

$sel:filter':ListAssessmentRuns', listAssessmentRuns_filter - You can use this parameter to specify a subset of data to be included in the action's response.

For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

$sel:assessmentTemplateArns:ListAssessmentRuns', listAssessmentRuns_assessmentTemplateArns - The ARNs that specify the assessment templates whose assessment runs you want to list.

$sel:maxResults:ListAssessmentRuns', listAssessmentRuns_maxResults - You can use this parameter to indicate the maximum number of items that you want in the response. The default value is 10. The maximum value is 500.

Request Lenses

listAssessmentRuns_nextToken :: Lens' ListAssessmentRuns (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListAssessmentRuns action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

listAssessmentRuns_filter :: Lens' ListAssessmentRuns (Maybe AssessmentRunFilter)

You can use this parameter to specify a subset of data to be included in the action's response.

For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

listAssessmentRuns_assessmentTemplateArns :: Lens' ListAssessmentRuns (Maybe [Text])

The ARNs that specify the assessment templates whose assessment runs you want to list.

listAssessmentRuns_maxResults :: Lens' ListAssessmentRuns (Maybe Int)

You can use this parameter to indicate the maximum number of items that you want in the response. The default value is 10. The maximum value is 500.

Destructuring the Response

data ListAssessmentRunsResponse

See: newListAssessmentRunsResponse smart constructor.

Constructors

ListAssessmentRunsResponse'

Fields

nextToken :: Maybe Text

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

httpStatus :: Int

The response's http status code.

assessmentRunArns :: [Text]

A list of ARNs that specifies the assessment runs that are returned by the action.

newListAssessmentRunsResponse

Arguments

:: Int
$sel:httpStatus:ListAssessmentRunsResponse'

-> ListAssessmentRunsResponse

Create a value of ListAssessmentRunsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAssessmentRuns', listAssessmentRunsResponse_nextToken - When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

$sel:httpStatus:ListAssessmentRunsResponse', listAssessmentRunsResponse_httpStatus - The response's http status code.

$sel:assessmentRunArns:ListAssessmentRunsResponse', listAssessmentRunsResponse_assessmentRunArns - A list of ARNs that specifies the assessment runs that are returned by the action.

Response Lenses

listAssessmentRunsResponse_nextToken :: Lens' ListAssessmentRunsResponse (Maybe Text)

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

listAssessmentRunsResponse_httpStatus :: Lens' ListAssessmentRunsResponse Int

The response's http status code.

listAssessmentRunsResponse_assessmentRunArns :: Lens' ListAssessmentRunsResponse [Text]

A list of ARNs that specifies the assessment runs that are returned by the action.