Network.AWS.Inspector.ListAssessmentTargets

Description

Lists the ARNs of the assessment targets within this AWS account. For more information about assessment targets, see Amazon Inspector Assessment Targets.

This operation returns paginated results.

Creating a Request

data ListAssessmentTargets

See: newListAssessmentTargets smart constructor.

Constructors

ListAssessmentTargets'

Fields

nextToken :: Maybe Text

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListAssessmentTargets action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

maxResults :: Maybe Int

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

filter' :: Maybe AssessmentTargetFilter

You can use this parameter to specify a subset of data to be included in the action's response.

For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

newListAssessmentTargets :: ListAssessmentTargets

Create a value of ListAssessmentTargets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAssessmentTargets', listAssessmentTargets_nextToken - You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListAssessmentTargets action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

$sel:maxResults:ListAssessmentTargets', listAssessmentTargets_maxResults - You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

$sel:filter':ListAssessmentTargets', listAssessmentTargets_filter - You can use this parameter to specify a subset of data to be included in the action's response.

For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

Request Lenses

listAssessmentTargets_nextToken :: Lens' ListAssessmentTargets (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListAssessmentTargets action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

listAssessmentTargets_maxResults :: Lens' ListAssessmentTargets (Maybe Int)

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

listAssessmentTargets_filter :: Lens' ListAssessmentTargets (Maybe AssessmentTargetFilter)

You can use this parameter to specify a subset of data to be included in the action's response.

For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

Destructuring the Response

data ListAssessmentTargetsResponse

See: newListAssessmentTargetsResponse smart constructor.

Constructors

ListAssessmentTargetsResponse'

Fields

nextToken :: Maybe Text

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

httpStatus :: Int

The response's http status code.

assessmentTargetArns :: [Text]

A list of ARNs that specifies the assessment targets that are returned by the action.

newListAssessmentTargetsResponse

Arguments

:: Int
$sel:httpStatus:ListAssessmentTargetsResponse'

-> ListAssessmentTargetsResponse

Create a value of ListAssessmentTargetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAssessmentTargets', listAssessmentTargetsResponse_nextToken - When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

$sel:httpStatus:ListAssessmentTargetsResponse', listAssessmentTargetsResponse_httpStatus - The response's http status code.

$sel:assessmentTargetArns:ListAssessmentTargetsResponse', listAssessmentTargetsResponse_assessmentTargetArns - A list of ARNs that specifies the assessment targets that are returned by the action.

Response Lenses

listAssessmentTargetsResponse_nextToken :: Lens' ListAssessmentTargetsResponse (Maybe Text)

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

listAssessmentTargetsResponse_httpStatus :: Lens' ListAssessmentTargetsResponse Int

The response's http status code.

listAssessmentTargetsResponse_assessmentTargetArns :: Lens' ListAssessmentTargetsResponse [Text]

A list of ARNs that specifies the assessment targets that are returned by the action.