Network.AWS.Inspector.ListEventSubscriptions

Description

Lists all the event subscriptions for the assessment template that is specified by the ARN of the assessment template. For more information, see SubscribeToEvent and UnsubscribeFromEvent.

This operation returns paginated results.

Creating a Request

data ListEventSubscriptions

See: newListEventSubscriptions smart constructor.

Constructors

ListEventSubscriptions'

Fields

resourceArn :: Maybe Text

The ARN of the assessment template for which you want to list the existing event subscriptions.

nextToken :: Maybe Text

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListEventSubscriptions action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

maxResults :: Maybe Int

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

newListEventSubscriptions :: ListEventSubscriptions

Create a value of ListEventSubscriptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListEventSubscriptions', listEventSubscriptions_resourceArn - The ARN of the assessment template for which you want to list the existing event subscriptions.

$sel:nextToken:ListEventSubscriptions', listEventSubscriptions_nextToken - You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListEventSubscriptions action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

$sel:maxResults:ListEventSubscriptions', listEventSubscriptions_maxResults - You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

Request Lenses

listEventSubscriptions_resourceArn :: Lens' ListEventSubscriptions (Maybe Text)

The ARN of the assessment template for which you want to list the existing event subscriptions.

listEventSubscriptions_nextToken :: Lens' ListEventSubscriptions (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListEventSubscriptions action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

listEventSubscriptions_maxResults :: Lens' ListEventSubscriptions (Maybe Int)

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

Destructuring the Response

data ListEventSubscriptionsResponse

See: newListEventSubscriptionsResponse smart constructor.

Constructors

ListEventSubscriptionsResponse'

Fields

nextToken :: Maybe Text

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

httpStatus :: Int

The response's http status code.

subscriptions :: [Subscription]

Details of the returned event subscriptions.

newListEventSubscriptionsResponse

Arguments

:: Int
$sel:httpStatus:ListEventSubscriptionsResponse'

-> ListEventSubscriptionsResponse

Create a value of ListEventSubscriptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEventSubscriptions', listEventSubscriptionsResponse_nextToken - When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

$sel:httpStatus:ListEventSubscriptionsResponse', listEventSubscriptionsResponse_httpStatus - The response's http status code.

$sel:subscriptions:ListEventSubscriptionsResponse', listEventSubscriptionsResponse_subscriptions - Details of the returned event subscriptions.

Response Lenses

listEventSubscriptionsResponse_nextToken :: Lens' ListEventSubscriptionsResponse (Maybe Text)

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

listEventSubscriptionsResponse_httpStatus :: Lens' ListEventSubscriptionsResponse Int

The response's http status code.

listEventSubscriptionsResponse_subscriptions :: Lens' ListEventSubscriptionsResponse [Subscription]

Details of the returned event subscriptions.