Network.AWS.Inspector.ListFindings

Description

Lists findings that are generated by the assessment runs that are specified by the ARNs of the assessment runs.

This operation returns paginated results.

Creating a Request

data ListFindings

See: newListFindings smart constructor.

Constructors

ListFindings'

Fields

nextToken :: Maybe Text

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListFindings action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

maxResults :: Maybe Int

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

assessmentRunArns :: Maybe [Text]

The ARNs of the assessment runs that generate the findings that you want to list.

filter' :: Maybe FindingFilter

You can use this parameter to specify a subset of data to be included in the action's response.

For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

newListFindings :: ListFindings

Create a value of ListFindings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFindings', listFindings_nextToken - You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListFindings action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

$sel:maxResults:ListFindings', listFindings_maxResults - You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

$sel:assessmentRunArns:ListFindings', listFindings_assessmentRunArns - The ARNs of the assessment runs that generate the findings that you want to list.

$sel:filter':ListFindings', listFindings_filter - You can use this parameter to specify a subset of data to be included in the action's response.

For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

Request Lenses

listFindings_nextToken :: Lens' ListFindings (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListFindings action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

listFindings_maxResults :: Lens' ListFindings (Maybe Int)

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

listFindings_assessmentRunArns :: Lens' ListFindings (Maybe [Text])

The ARNs of the assessment runs that generate the findings that you want to list.

listFindings_filter :: Lens' ListFindings (Maybe FindingFilter)

You can use this parameter to specify a subset of data to be included in the action's response.

For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

Destructuring the Response

data ListFindingsResponse

See: newListFindingsResponse smart constructor.

Constructors

ListFindingsResponse'

Fields

nextToken :: Maybe Text

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

httpStatus :: Int

The response's http status code.

findingArns :: [Text]

A list of ARNs that specifies the findings returned by the action.

newListFindingsResponse

Arguments

:: Int
$sel:httpStatus:ListFindingsResponse'

-> ListFindingsResponse

Create a value of ListFindingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFindings', listFindingsResponse_nextToken - When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

$sel:httpStatus:ListFindingsResponse', listFindingsResponse_httpStatus - The response's http status code.

$sel:findingArns:ListFindingsResponse', listFindingsResponse_findingArns - A list of ARNs that specifies the findings returned by the action.

Response Lenses

listFindingsResponse_nextToken :: Lens' ListFindingsResponse (Maybe Text)

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

listFindingsResponse_httpStatus :: Lens' ListFindingsResponse Int

The response's http status code.

listFindingsResponse_findingArns :: Lens' ListFindingsResponse [Text]

A list of ARNs that specifies the findings returned by the action.