Network.AWS.Inspector.RemoveAttributesFromFindings

Description

Removes entire attributes (key and value pairs) from the findings that are specified by the ARNs of the findings where an attribute with the specified key exists.

Creating a Request

data RemoveAttributesFromFindings

See: newRemoveAttributesFromFindings smart constructor.

Constructors

RemoveAttributesFromFindings'

Fields

findingArns :: NonEmpty Text

The ARNs that specify the findings that you want to remove attributes from.

attributeKeys :: [Text]

The array of attribute keys that you want to remove from specified findings.

newRemoveAttributesFromFindings

Arguments

:: NonEmpty Text
$sel:findingArns:RemoveAttributesFromFindings'

-> RemoveAttributesFromFindings

Create a value of RemoveAttributesFromFindings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:findingArns:RemoveAttributesFromFindings', removeAttributesFromFindings_findingArns - The ARNs that specify the findings that you want to remove attributes from.

$sel:attributeKeys:RemoveAttributesFromFindings', removeAttributesFromFindings_attributeKeys - The array of attribute keys that you want to remove from specified findings.

Request Lenses

removeAttributesFromFindings_findingArns :: Lens' RemoveAttributesFromFindings (NonEmpty Text)

The ARNs that specify the findings that you want to remove attributes from.

removeAttributesFromFindings_attributeKeys :: Lens' RemoveAttributesFromFindings [Text]

The array of attribute keys that you want to remove from specified findings.

Destructuring the Response

data RemoveAttributesFromFindingsResponse

See: newRemoveAttributesFromFindingsResponse smart constructor.

Constructors

RemoveAttributesFromFindingsResponse'

Fields

httpStatus :: Int

The response's http status code.

failedItems :: HashMap Text FailedItemDetails

Attributes details that cannot be described. An error code is provided for each failed item.

newRemoveAttributesFromFindingsResponse

Arguments

:: Int
$sel:httpStatus:RemoveAttributesFromFindingsResponse'

-> RemoveAttributesFromFindingsResponse

Create a value of RemoveAttributesFromFindingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:RemoveAttributesFromFindingsResponse', removeAttributesFromFindingsResponse_httpStatus - The response's http status code.

$sel:failedItems:RemoveAttributesFromFindingsResponse', removeAttributesFromFindingsResponse_failedItems - Attributes details that cannot be described. An error code is provided for each failed item.

Response Lenses

removeAttributesFromFindingsResponse_httpStatus :: Lens' RemoveAttributesFromFindingsResponse Int

The response's http status code.

removeAttributesFromFindingsResponse_failedItems :: Lens' RemoveAttributesFromFindingsResponse (HashMap Text FailedItemDetails)

Attributes details that cannot be described. An error code is provided for each failed item.