Network.AWS.Inspector.SetTagsForResource

Description

Sets tags (key and value pairs) to the assessment template that is specified by the ARN of the assessment template.

Creating a Request

data SetTagsForResource

See: newSetTagsForResource smart constructor.

Constructors

SetTagsForResource'

Fields

tags :: Maybe [Tag]

A collection of key and value pairs that you want to set to the assessment template.

resourceArn :: Text

The ARN of the assessment template that you want to set tags to.

newSetTagsForResource

Arguments

:: Text
$sel:resourceArn:SetTagsForResource'

-> SetTagsForResource

Create a value of SetTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:SetTagsForResource', setTagsForResource_tags - A collection of key and value pairs that you want to set to the assessment template.

$sel:resourceArn:SetTagsForResource', setTagsForResource_resourceArn - The ARN of the assessment template that you want to set tags to.

Request Lenses

setTagsForResource_tags :: Lens' SetTagsForResource (Maybe [Tag])

A collection of key and value pairs that you want to set to the assessment template.

setTagsForResource_resourceArn :: Lens' SetTagsForResource Text

The ARN of the assessment template that you want to set tags to.

Destructuring the Response

data SetTagsForResourceResponse

See: newSetTagsForResourceResponse smart constructor.

Constructors

SetTagsForResourceResponse'

newSetTagsForResourceResponse :: SetTagsForResourceResponse

Create a value of SetTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.