Network.AWS.Inspector.Types.AgentPreview

Description

Documentation

data AgentPreview

Used as a response element in the PreviewAgents action.

See: newAgentPreview smart constructor.

Constructors

AgentPreview'

Fields

hostname :: Maybe Text

The hostname of the EC2 instance on which the Amazon Inspector Agent is installed.

agentVersion :: Maybe Text

The version of the Amazon Inspector Agent.

kernelVersion :: Maybe Text

The kernel version of the operating system running on the EC2 instance on which the Amazon Inspector Agent is installed.

operatingSystem :: Maybe Text

The operating system running on the EC2 instance on which the Amazon Inspector Agent is installed.

agentHealth :: Maybe AgentHealth

The health status of the Amazon Inspector Agent.

autoScalingGroup :: Maybe Text

The Auto Scaling group for the EC2 instance where the agent is installed.

ipv4Address :: Maybe Text

The IP address of the EC2 instance on which the Amazon Inspector Agent is installed.

agentId :: Text

The ID of the EC2 instance where the agent is installed.

newAgentPreview

Arguments

:: Text
$sel:agentId:AgentPreview'

-> AgentPreview

Create a value of AgentPreview with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostname:AgentPreview', agentPreview_hostname - The hostname of the EC2 instance on which the Amazon Inspector Agent is installed.

$sel:agentVersion:AgentPreview', agentPreview_agentVersion - The version of the Amazon Inspector Agent.

$sel:kernelVersion:AgentPreview', agentPreview_kernelVersion - The kernel version of the operating system running on the EC2 instance on which the Amazon Inspector Agent is installed.

$sel:operatingSystem:AgentPreview', agentPreview_operatingSystem - The operating system running on the EC2 instance on which the Amazon Inspector Agent is installed.

$sel:agentHealth:AgentPreview', agentPreview_agentHealth - The health status of the Amazon Inspector Agent.

$sel:autoScalingGroup:AgentPreview', agentPreview_autoScalingGroup - The Auto Scaling group for the EC2 instance where the agent is installed.

$sel:ipv4Address:AgentPreview', agentPreview_ipv4Address - The IP address of the EC2 instance on which the Amazon Inspector Agent is installed.

$sel:agentId:AgentPreview', agentPreview_agentId - The ID of the EC2 instance where the agent is installed.

agentPreview_hostname :: Lens' AgentPreview (Maybe Text)

The hostname of the EC2 instance on which the Amazon Inspector Agent is installed.

agentPreview_agentVersion :: Lens' AgentPreview (Maybe Text)

The version of the Amazon Inspector Agent.

agentPreview_kernelVersion :: Lens' AgentPreview (Maybe Text)

The kernel version of the operating system running on the EC2 instance on which the Amazon Inspector Agent is installed.

agentPreview_operatingSystem :: Lens' AgentPreview (Maybe Text)

The operating system running on the EC2 instance on which the Amazon Inspector Agent is installed.

agentPreview_agentHealth :: Lens' AgentPreview (Maybe AgentHealth)

The health status of the Amazon Inspector Agent.

agentPreview_autoScalingGroup :: Lens' AgentPreview (Maybe Text)

The Auto Scaling group for the EC2 instance where the agent is installed.

agentPreview_ipv4Address :: Lens' AgentPreview (Maybe Text)

The IP address of the EC2 instance on which the Amazon Inspector Agent is installed.

agentPreview_agentId :: Lens' AgentPreview Text

The ID of the EC2 instance where the agent is installed.