Network.AWS.Inspector.Types.AssessmentRunNotification

Description

Documentation

data AssessmentRunNotification

Used as one of the elements of the AssessmentRun data type.

See: newAssessmentRunNotification smart constructor.

Constructors

AssessmentRunNotification'

Fields

snsTopicArn :: Maybe Text

The SNS topic to which the SNS notification is sent.

snsPublishStatusCode :: Maybe AssessmentRunNotificationSnsStatusCode

The status code of the SNS notification.

message :: Maybe Text

The message included in the notification.

date :: POSIX

The date of the notification.

event :: InspectorEvent

The event for which a notification is sent.

error :: Bool

The Boolean value that specifies whether the notification represents an error.

newAssessmentRunNotification

Arguments

:: UTCTime
$sel:date:AssessmentRunNotification'

-> InspectorEvent
$sel:event:AssessmentRunNotification'

-> Bool
$sel:error:AssessmentRunNotification'

-> AssessmentRunNotification

Create a value of AssessmentRunNotification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snsTopicArn:AssessmentRunNotification', assessmentRunNotification_snsTopicArn - The SNS topic to which the SNS notification is sent.

$sel:snsPublishStatusCode:AssessmentRunNotification', assessmentRunNotification_snsPublishStatusCode - The status code of the SNS notification.

$sel:message:AssessmentRunNotification', assessmentRunNotification_message - The message included in the notification.

$sel:date:AssessmentRunNotification', assessmentRunNotification_date - The date of the notification.

$sel:event:AssessmentRunNotification', assessmentRunNotification_event - The event for which a notification is sent.

$sel:error:AssessmentRunNotification', assessmentRunNotification_error - The Boolean value that specifies whether the notification represents an error.

assessmentRunNotification_snsTopicArn :: Lens' AssessmentRunNotification (Maybe Text)

The SNS topic to which the SNS notification is sent.

assessmentRunNotification_snsPublishStatusCode :: Lens' AssessmentRunNotification (Maybe AssessmentRunNotificationSnsStatusCode)

The status code of the SNS notification.

assessmentRunNotification_message :: Lens' AssessmentRunNotification (Maybe Text)

The message included in the notification.

assessmentRunNotification_date :: Lens' AssessmentRunNotification UTCTime

The date of the notification.

assessmentRunNotification_event :: Lens' AssessmentRunNotification InspectorEvent

The event for which a notification is sent.

assessmentRunNotification_error :: Lens' AssessmentRunNotification Bool

The Boolean value that specifies whether the notification represents an error.