Network.AWS.Inspector.Types.AssessmentRunStateChange

Description

Documentation

data AssessmentRunStateChange

Used as one of the elements of the AssessmentRun data type.

See: newAssessmentRunStateChange smart constructor.

Constructors

AssessmentRunStateChange'

Fields

stateChangedAt :: POSIX

The last time the assessment run state changed.

state :: AssessmentRunState

The assessment run state.

newAssessmentRunStateChange

Arguments

:: UTCTime
$sel:stateChangedAt:AssessmentRunStateChange'

-> AssessmentRunState
$sel:state:AssessmentRunStateChange'

-> AssessmentRunStateChange

Create a value of AssessmentRunStateChange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stateChangedAt:AssessmentRunStateChange', assessmentRunStateChange_stateChangedAt - The last time the assessment run state changed.

$sel:state:AssessmentRunStateChange', assessmentRunStateChange_state - The assessment run state.

assessmentRunStateChange_stateChangedAt :: Lens' AssessmentRunStateChange UTCTime

The last time the assessment run state changed.

assessmentRunStateChange_state :: Lens' AssessmentRunStateChange AssessmentRunState

The assessment run state.