Network.AWS.Inspector.Types.AssessmentTemplateFilter

Description

Documentation

data AssessmentTemplateFilter

Used as the request parameter in the ListAssessmentTemplates action.

See: newAssessmentTemplateFilter smart constructor.

Constructors

AssessmentTemplateFilter'

Fields

durationRange :: Maybe DurationRange

For a record to match a filter, the value specified for this data type property must inclusively match any value between the specified minimum and maximum values of the durationInSeconds property of the AssessmentTemplate data type.

rulesPackageArns :: Maybe [Text]

For a record to match a filter, the values that are specified for this data type property must be contained in the list of values of the rulesPackageArns property of the AssessmentTemplate data type.

namePattern :: Maybe Text

For a record to match a filter, an explicit value or a string that contains a wildcard that is specified for this data type property must match the value of the assessmentTemplateName property of the AssessmentTemplate data type.

newAssessmentTemplateFilter :: AssessmentTemplateFilter

Create a value of AssessmentTemplateFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:durationRange:AssessmentTemplateFilter', assessmentTemplateFilter_durationRange - For a record to match a filter, the value specified for this data type property must inclusively match any value between the specified minimum and maximum values of the durationInSeconds property of the AssessmentTemplate data type.

$sel:rulesPackageArns:AssessmentTemplateFilter', assessmentTemplateFilter_rulesPackageArns - For a record to match a filter, the values that are specified for this data type property must be contained in the list of values of the rulesPackageArns property of the AssessmentTemplate data type.

$sel:namePattern:AssessmentTemplateFilter', assessmentTemplateFilter_namePattern - For a record to match a filter, an explicit value or a string that contains a wildcard that is specified for this data type property must match the value of the assessmentTemplateName property of the AssessmentTemplate data type.

assessmentTemplateFilter_durationRange :: Lens' AssessmentTemplateFilter (Maybe DurationRange)

For a record to match a filter, the value specified for this data type property must inclusively match any value between the specified minimum and maximum values of the durationInSeconds property of the AssessmentTemplate data type.

assessmentTemplateFilter_rulesPackageArns :: Lens' AssessmentTemplateFilter (Maybe [Text])

For a record to match a filter, the values that are specified for this data type property must be contained in the list of values of the rulesPackageArns property of the AssessmentTemplate data type.

assessmentTemplateFilter_namePattern :: Lens' AssessmentTemplateFilter (Maybe Text)

For a record to match a filter, an explicit value or a string that contains a wildcard that is specified for this data type property must match the value of the assessmentTemplateName property of the AssessmentTemplate data type.