Network.AWS.Inspector.Types.AssetAttributes

Description

Documentation

data AssetAttributes

A collection of attributes of the host from which the finding is generated.

See: newAssetAttributes smart constructor.

Constructors

AssetAttributes'

Fields

hostname :: Maybe Text

The hostname of the EC2 instance where the finding is generated.

agentId :: Maybe Text

The ID of the agent that is installed on the EC2 instance where the finding is generated.

amiId :: Maybe Text

The ID of the Amazon Machine Image (AMI) that is installed on the EC2 instance where the finding is generated.

tags :: Maybe [Tag]

The tags related to the EC2 instance where the finding is generated.

ipv4Addresses :: Maybe [Text]

The list of IP v4 addresses of the EC2 instance where the finding is generated.

networkInterfaces :: Maybe [NetworkInterface]

An array of the network interfaces interacting with the EC2 instance where the finding is generated.

autoScalingGroup :: Maybe Text

The Auto Scaling group of the EC2 instance where the finding is generated.

schemaVersion :: Natural

The schema version of this data type.

newAssetAttributes

Arguments

:: Natural
$sel:schemaVersion:AssetAttributes'

-> AssetAttributes

Create a value of AssetAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostname:AssetAttributes', assetAttributes_hostname - The hostname of the EC2 instance where the finding is generated.

$sel:agentId:AssetAttributes', assetAttributes_agentId - The ID of the agent that is installed on the EC2 instance where the finding is generated.

$sel:amiId:AssetAttributes', assetAttributes_amiId - The ID of the Amazon Machine Image (AMI) that is installed on the EC2 instance where the finding is generated.

$sel:tags:AssetAttributes', assetAttributes_tags - The tags related to the EC2 instance where the finding is generated.

$sel:ipv4Addresses:AssetAttributes', assetAttributes_ipv4Addresses - The list of IP v4 addresses of the EC2 instance where the finding is generated.

$sel:networkInterfaces:AssetAttributes', assetAttributes_networkInterfaces - An array of the network interfaces interacting with the EC2 instance where the finding is generated.

$sel:autoScalingGroup:AssetAttributes', assetAttributes_autoScalingGroup - The Auto Scaling group of the EC2 instance where the finding is generated.

$sel:schemaVersion:AssetAttributes', assetAttributes_schemaVersion - The schema version of this data type.

assetAttributes_hostname :: Lens' AssetAttributes (Maybe Text)

The hostname of the EC2 instance where the finding is generated.

assetAttributes_agentId :: Lens' AssetAttributes (Maybe Text)

The ID of the agent that is installed on the EC2 instance where the finding is generated.

assetAttributes_amiId :: Lens' AssetAttributes (Maybe Text)

The ID of the Amazon Machine Image (AMI) that is installed on the EC2 instance where the finding is generated.

assetAttributes_tags :: Lens' AssetAttributes (Maybe [Tag])

The tags related to the EC2 instance where the finding is generated.

assetAttributes_ipv4Addresses :: Lens' AssetAttributes (Maybe [Text])

The list of IP v4 addresses of the EC2 instance where the finding is generated.

assetAttributes_networkInterfaces :: Lens' AssetAttributes (Maybe [NetworkInterface])

An array of the network interfaces interacting with the EC2 instance where the finding is generated.

assetAttributes_autoScalingGroup :: Lens' AssetAttributes (Maybe Text)

The Auto Scaling group of the EC2 instance where the finding is generated.

assetAttributes_schemaVersion :: Lens' AssetAttributes Natural

The schema version of this data type.