Network.AWS.Inspector.Types.Exclusion

Description

Documentation

data Exclusion

Contains information about what was excluded from an assessment run.

See: newExclusion smart constructor.

Constructors

Exclusion'

Fields

attributes :: Maybe [Attribute]

The system-defined attributes for the exclusion.

arn :: Text

The ARN that specifies the exclusion.

title :: Text

The name of the exclusion.

description :: Text

The description of the exclusion.

recommendation :: Text

The recommendation for the exclusion.

scopes :: NonEmpty Scope

The AWS resources for which the exclusion pertains.

newExclusion

Arguments

:: Text
$sel:arn:Exclusion'

-> Text
$sel:title:Exclusion'

-> Text
$sel:description:Exclusion'

-> Text
$sel:recommendation:Exclusion'

-> NonEmpty Scope
$sel:scopes:Exclusion'

-> Exclusion

Create a value of Exclusion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributes:Exclusion', exclusion_attributes - The system-defined attributes for the exclusion.

$sel:arn:Exclusion', exclusion_arn - The ARN that specifies the exclusion.

$sel:title:Exclusion', exclusion_title - The name of the exclusion.

$sel:description:Exclusion', exclusion_description - The description of the exclusion.

$sel:recommendation:Exclusion', exclusion_recommendation - The recommendation for the exclusion.

$sel:scopes:Exclusion', exclusion_scopes - The AWS resources for which the exclusion pertains.

exclusion_attributes :: Lens' Exclusion (Maybe [Attribute])

The system-defined attributes for the exclusion.

exclusion_arn :: Lens' Exclusion Text

The ARN that specifies the exclusion.

exclusion_title :: Lens' Exclusion Text

The name of the exclusion.

exclusion_description :: Lens' Exclusion Text

The description of the exclusion.

exclusion_recommendation :: Lens' Exclusion Text

The recommendation for the exclusion.

exclusion_scopes :: Lens' Exclusion (NonEmpty Scope)

The AWS resources for which the exclusion pertains.