Network.AWS.Inspector.Types.FindingFilter

Description

Documentation

data FindingFilter

This data type is used as a request parameter in the ListFindings action.

See: newFindingFilter smart constructor.

Constructors

FindingFilter'

Fields

agentIds :: Maybe [Text]

For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the agentId property of the Finding data type.

ruleNames :: Maybe [Text]

For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the ruleName property of the Finding data type.

userAttributes :: Maybe [Attribute]

For a record to match a filter, the value that is specified for this data type property must be contained in the list of values of the userAttributes property of the Finding data type.

rulesPackageArns :: Maybe [Text]

For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the rulesPackageArn property of the Finding data type.

attributes :: Maybe [Attribute]

For a record to match a filter, the list of values that are specified for this data type property must be contained in the list of values of the attributes property of the Finding data type.

severities :: Maybe [Severity]

For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the severity property of the Finding data type.

creationTimeRange :: Maybe TimestampRange

The time range during which the finding is generated.

autoScalingGroups :: Maybe [Text]

For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the autoScalingGroup property of the Finding data type.

newFindingFilter :: FindingFilter

Create a value of FindingFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:agentIds:FindingFilter', findingFilter_agentIds - For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the agentId property of the Finding data type.

$sel:ruleNames:FindingFilter', findingFilter_ruleNames - For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the ruleName property of the Finding data type.

$sel:userAttributes:FindingFilter', findingFilter_userAttributes - For a record to match a filter, the value that is specified for this data type property must be contained in the list of values of the userAttributes property of the Finding data type.

$sel:rulesPackageArns:FindingFilter', findingFilter_rulesPackageArns - For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the rulesPackageArn property of the Finding data type.

$sel:attributes:FindingFilter', findingFilter_attributes - For a record to match a filter, the list of values that are specified for this data type property must be contained in the list of values of the attributes property of the Finding data type.

$sel:severities:FindingFilter', findingFilter_severities - For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the severity property of the Finding data type.

$sel:creationTimeRange:FindingFilter', findingFilter_creationTimeRange - The time range during which the finding is generated.

$sel:autoScalingGroups:FindingFilter', findingFilter_autoScalingGroups - For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the autoScalingGroup property of the Finding data type.

findingFilter_agentIds :: Lens' FindingFilter (Maybe [Text])

For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the agentId property of the Finding data type.

findingFilter_ruleNames :: Lens' FindingFilter (Maybe [Text])

For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the ruleName property of the Finding data type.

findingFilter_userAttributes :: Lens' FindingFilter (Maybe [Attribute])

For a record to match a filter, the value that is specified for this data type property must be contained in the list of values of the userAttributes property of the Finding data type.

findingFilter_rulesPackageArns :: Lens' FindingFilter (Maybe [Text])

For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the rulesPackageArn property of the Finding data type.

findingFilter_attributes :: Lens' FindingFilter (Maybe [Attribute])

For a record to match a filter, the list of values that are specified for this data type property must be contained in the list of values of the attributes property of the Finding data type.

findingFilter_severities :: Lens' FindingFilter (Maybe [Severity])

For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the severity property of the Finding data type.

findingFilter_creationTimeRange :: Lens' FindingFilter (Maybe TimestampRange)

The time range during which the finding is generated.

findingFilter_autoScalingGroups :: Lens' FindingFilter (Maybe [Text])

For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the autoScalingGroup property of the Finding data type.