Network.AWS.Inspector.Types.InspectorServiceAttributes

Description

Documentation

data InspectorServiceAttributes

This data type is used in the Finding data type.

See: newInspectorServiceAttributes smart constructor.

Constructors

InspectorServiceAttributes'

Fields

rulesPackageArn :: Maybe Text

The ARN of the rules package that is used to generate the finding.

assessmentRunArn :: Maybe Text

The ARN of the assessment run during which the finding is generated.

schemaVersion :: Natural

The schema version of this data type.

newInspectorServiceAttributes

Arguments

:: Natural
$sel:schemaVersion:InspectorServiceAttributes'

-> InspectorServiceAttributes

Create a value of InspectorServiceAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rulesPackageArn:InspectorServiceAttributes', inspectorServiceAttributes_rulesPackageArn - The ARN of the rules package that is used to generate the finding.

$sel:assessmentRunArn:InspectorServiceAttributes', inspectorServiceAttributes_assessmentRunArn - The ARN of the assessment run during which the finding is generated.

$sel:schemaVersion:InspectorServiceAttributes', inspectorServiceAttributes_schemaVersion - The schema version of this data type.

inspectorServiceAttributes_rulesPackageArn :: Lens' InspectorServiceAttributes (Maybe Text)

The ARN of the rules package that is used to generate the finding.

inspectorServiceAttributes_assessmentRunArn :: Lens' InspectorServiceAttributes (Maybe Text)

The ARN of the assessment run during which the finding is generated.

inspectorServiceAttributes_schemaVersion :: Lens' InspectorServiceAttributes Natural

The schema version of this data type.