Network.AWS.Inspector.Types.NetworkInterface

Description

Documentation

data NetworkInterface

Contains information about the network interfaces interacting with an EC2 instance. This data type is used as one of the elements of the AssetAttributes data type.

See: newNetworkInterface smart constructor.

Constructors

NetworkInterface'

Fields

privateIpAddresses :: Maybe [PrivateIp]

A list of the private IP addresses associated with the network interface. Includes the privateDnsName and privateIpAddress.

ipv6Addresses :: Maybe [Text]

The IP addresses associated with the network interface.

securityGroups :: Maybe [SecurityGroup]

A list of the security groups associated with the network interface. Includes the groupId and groupName.

publicDnsName :: Maybe Text

The name of a public DNS associated with the network interface.

subnetId :: Maybe Text

The ID of a subnet associated with the network interface.

networkInterfaceId :: Maybe Text

The ID of the network interface.

privateDnsName :: Maybe Text

The name of a private DNS associated with the network interface.

vpcId :: Maybe Text

The ID of a VPC associated with the network interface.

publicIp :: Maybe Text

The public IP address from which the network interface is reachable.

privateIpAddress :: Maybe Text

The private IP address associated with the network interface.

newNetworkInterface :: NetworkInterface

Create a value of NetworkInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:privateIpAddresses:NetworkInterface', networkInterface_privateIpAddresses - A list of the private IP addresses associated with the network interface. Includes the privateDnsName and privateIpAddress.

$sel:ipv6Addresses:NetworkInterface', networkInterface_ipv6Addresses - The IP addresses associated with the network interface.

$sel:securityGroups:NetworkInterface', networkInterface_securityGroups - A list of the security groups associated with the network interface. Includes the groupId and groupName.

$sel:publicDnsName:NetworkInterface', networkInterface_publicDnsName - The name of a public DNS associated with the network interface.

$sel:subnetId:NetworkInterface', networkInterface_subnetId - The ID of a subnet associated with the network interface.

$sel:networkInterfaceId:NetworkInterface', networkInterface_networkInterfaceId - The ID of the network interface.

$sel:privateDnsName:NetworkInterface', networkInterface_privateDnsName - The name of a private DNS associated with the network interface.

$sel:vpcId:NetworkInterface', networkInterface_vpcId - The ID of a VPC associated with the network interface.

$sel:publicIp:NetworkInterface', networkInterface_publicIp - The public IP address from which the network interface is reachable.

$sel:privateIpAddress:NetworkInterface', networkInterface_privateIpAddress - The private IP address associated with the network interface.

networkInterface_privateIpAddresses :: Lens' NetworkInterface (Maybe [PrivateIp])

A list of the private IP addresses associated with the network interface. Includes the privateDnsName and privateIpAddress.

networkInterface_ipv6Addresses :: Lens' NetworkInterface (Maybe [Text])

The IP addresses associated with the network interface.

networkInterface_securityGroups :: Lens' NetworkInterface (Maybe [SecurityGroup])

A list of the security groups associated with the network interface. Includes the groupId and groupName.

networkInterface_publicDnsName :: Lens' NetworkInterface (Maybe Text)

The name of a public DNS associated with the network interface.

networkInterface_subnetId :: Lens' NetworkInterface (Maybe Text)

The ID of a subnet associated with the network interface.

networkInterface_networkInterfaceId :: Lens' NetworkInterface (Maybe Text)

The ID of the network interface.

networkInterface_privateDnsName :: Lens' NetworkInterface (Maybe Text)

The name of a private DNS associated with the network interface.

networkInterface_vpcId :: Lens' NetworkInterface (Maybe Text)

The ID of a VPC associated with the network interface.

networkInterface_publicIp :: Lens' NetworkInterface (Maybe Text)

The public IP address from which the network interface is reachable.

networkInterface_privateIpAddress :: Lens' NetworkInterface (Maybe Text)

The private IP address associated with the network interface.