Network.AWS.Inspector.Types.Scope

Description

Documentation

data Scope

This data type contains key-value pairs that identify various Amazon resources.

See: newScope smart constructor.

Constructors

Scope'

Fields

value :: Maybe Text

The resource identifier for the specified scope type.

key :: Maybe ScopeType

The type of the scope.

newScope :: Scope

Create a value of Scope with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:Scope', scope_value - The resource identifier for the specified scope type.

$sel:key:Scope', scope_key - The type of the scope.

scope_value :: Lens' Scope (Maybe Text)

The resource identifier for the specified scope type.

scope_key :: Lens' Scope (Maybe ScopeType)

The type of the scope.