Network.AWS.IoT.AddThingToBillingGroup

Description

Adds a thing to a billing group.

Requires permission to access the AddThingToBillingGroup action.

Creating a Request

data AddThingToBillingGroup

See: newAddThingToBillingGroup smart constructor.

Constructors

AddThingToBillingGroup'

Fields

thingArn :: Maybe Text

The ARN of the thing to be added to the billing group.

billingGroupArn :: Maybe Text

The ARN of the billing group.

thingName :: Maybe Text

The name of the thing to be added to the billing group.

billingGroupName :: Maybe Text

The name of the billing group.

newAddThingToBillingGroup :: AddThingToBillingGroup

Create a value of AddThingToBillingGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thingArn:AddThingToBillingGroup', addThingToBillingGroup_thingArn - The ARN of the thing to be added to the billing group.

$sel:billingGroupArn:AddThingToBillingGroup', addThingToBillingGroup_billingGroupArn - The ARN of the billing group.

$sel:thingName:AddThingToBillingGroup', addThingToBillingGroup_thingName - The name of the thing to be added to the billing group.

$sel:billingGroupName:AddThingToBillingGroup', addThingToBillingGroup_billingGroupName - The name of the billing group.

Request Lenses

addThingToBillingGroup_thingArn :: Lens' AddThingToBillingGroup (Maybe Text)

The ARN of the thing to be added to the billing group.

addThingToBillingGroup_billingGroupArn :: Lens' AddThingToBillingGroup (Maybe Text)

The ARN of the billing group.

addThingToBillingGroup_thingName :: Lens' AddThingToBillingGroup (Maybe Text)

The name of the thing to be added to the billing group.

addThingToBillingGroup_billingGroupName :: Lens' AddThingToBillingGroup (Maybe Text)

The name of the billing group.

Destructuring the Response

data AddThingToBillingGroupResponse

See: newAddThingToBillingGroupResponse smart constructor.

Constructors

AddThingToBillingGroupResponse'

Fields

httpStatus :: Int

The response's http status code.

newAddThingToBillingGroupResponse

Arguments

:: Int
$sel:httpStatus:AddThingToBillingGroupResponse'

-> AddThingToBillingGroupResponse

Create a value of AddThingToBillingGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AddThingToBillingGroupResponse', addThingToBillingGroupResponse_httpStatus - The response's http status code.

Response Lenses

addThingToBillingGroupResponse_httpStatus :: Lens' AddThingToBillingGroupResponse Int

The response's http status code.