Network.AWS.IoT.AssociateTargetsWithJob

Description

Associates a group with a continuous job. The following criteria must be met:

The job must have been created with the targetSelection field set to "CONTINUOUS".

The job status must currently be "IN_PROGRESS".

The total number of targets associated with a job must not exceed 100.

Requires permission to access the AssociateTargetsWithJob action.

Creating a Request

data AssociateTargetsWithJob

See: newAssociateTargetsWithJob smart constructor.

Constructors

AssociateTargetsWithJob'

Fields

namespaceId :: Maybe Text

The namespace used to indicate that a job is a customer-managed job.

When you specify a value for this parameter, Amazon Web Services IoT Core sends jobs notifications to MQTT topics that contain the value in the following format.

$aws/things/THING_NAME/jobs/JOB_ID/notify-namespace-NAMESPACE_ID/

The namespaceId feature is in public preview.

comment :: Maybe Text

An optional comment string describing why the job was associated with the targets.

targets :: NonEmpty Text

A list of thing group ARNs that define the targets of the job.

jobId :: Text

The unique identifier you assigned to this job when it was created.

newAssociateTargetsWithJob

Arguments

:: NonEmpty Text
$sel:targets:AssociateTargetsWithJob'

-> Text
$sel:jobId:AssociateTargetsWithJob'

-> AssociateTargetsWithJob

Create a value of AssociateTargetsWithJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:namespaceId:AssociateTargetsWithJob', associateTargetsWithJob_namespaceId - The namespace used to indicate that a job is a customer-managed job.

When you specify a value for this parameter, Amazon Web Services IoT Core sends jobs notifications to MQTT topics that contain the value in the following format.

$aws/things/THING_NAME/jobs/JOB_ID/notify-namespace-NAMESPACE_ID/

The namespaceId feature is in public preview.

$sel:comment:AssociateTargetsWithJob', associateTargetsWithJob_comment - An optional comment string describing why the job was associated with the targets.

$sel:targets:AssociateTargetsWithJob', associateTargetsWithJob_targets - A list of thing group ARNs that define the targets of the job.

$sel:jobId:AssociateTargetsWithJob', associateTargetsWithJob_jobId - The unique identifier you assigned to this job when it was created.

Request Lenses

associateTargetsWithJob_namespaceId :: Lens' AssociateTargetsWithJob (Maybe Text)

The namespace used to indicate that a job is a customer-managed job.

When you specify a value for this parameter, Amazon Web Services IoT Core sends jobs notifications to MQTT topics that contain the value in the following format.

$aws/things/THING_NAME/jobs/JOB_ID/notify-namespace-NAMESPACE_ID/

The namespaceId feature is in public preview.

associateTargetsWithJob_comment :: Lens' AssociateTargetsWithJob (Maybe Text)

An optional comment string describing why the job was associated with the targets.

associateTargetsWithJob_targets :: Lens' AssociateTargetsWithJob (NonEmpty Text)

A list of thing group ARNs that define the targets of the job.

associateTargetsWithJob_jobId :: Lens' AssociateTargetsWithJob Text

The unique identifier you assigned to this job when it was created.

Destructuring the Response

data AssociateTargetsWithJobResponse

See: newAssociateTargetsWithJobResponse smart constructor.

Constructors

AssociateTargetsWithJobResponse'

Fields

jobArn :: Maybe Text

An ARN identifying the job.

description :: Maybe Text

A short text description of the job.

jobId :: Maybe Text

The unique identifier you assigned to this job when it was created.

httpStatus :: Int

The response's http status code.

newAssociateTargetsWithJobResponse

Arguments

:: Int
$sel:httpStatus:AssociateTargetsWithJobResponse'

-> AssociateTargetsWithJobResponse

Create a value of AssociateTargetsWithJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobArn:AssociateTargetsWithJobResponse', associateTargetsWithJobResponse_jobArn - An ARN identifying the job.

$sel:description:AssociateTargetsWithJobResponse', associateTargetsWithJobResponse_description - A short text description of the job.

$sel:jobId:AssociateTargetsWithJob', associateTargetsWithJobResponse_jobId - The unique identifier you assigned to this job when it was created.

$sel:httpStatus:AssociateTargetsWithJobResponse', associateTargetsWithJobResponse_httpStatus - The response's http status code.

Response Lenses

associateTargetsWithJobResponse_jobArn :: Lens' AssociateTargetsWithJobResponse (Maybe Text)

An ARN identifying the job.

associateTargetsWithJobResponse_description :: Lens' AssociateTargetsWithJobResponse (Maybe Text)

A short text description of the job.

associateTargetsWithJobResponse_jobId :: Lens' AssociateTargetsWithJobResponse (Maybe Text)

The unique identifier you assigned to this job when it was created.

associateTargetsWithJobResponse_httpStatus :: Lens' AssociateTargetsWithJobResponse Int

The response's http status code.