Network.AWS.IoT.CancelAuditMitigationActionsTask

Description

Cancels a mitigation action task that is in progress. If the task is not in progress, an InvalidRequestException occurs.

Requires permission to access the CancelAuditMitigationActionsTask action.

Creating a Request

data CancelAuditMitigationActionsTask

See: newCancelAuditMitigationActionsTask smart constructor.

Constructors

CancelAuditMitigationActionsTask'

Fields

taskId :: Text

The unique identifier for the task that you want to cancel.

newCancelAuditMitigationActionsTask

Arguments

:: Text
$sel:taskId:CancelAuditMitigationActionsTask'

-> CancelAuditMitigationActionsTask

Create a value of CancelAuditMitigationActionsTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskId:CancelAuditMitigationActionsTask', cancelAuditMitigationActionsTask_taskId - The unique identifier for the task that you want to cancel.

Request Lenses

cancelAuditMitigationActionsTask_taskId :: Lens' CancelAuditMitigationActionsTask Text

The unique identifier for the task that you want to cancel.

Destructuring the Response

data CancelAuditMitigationActionsTaskResponse

See: newCancelAuditMitigationActionsTaskResponse smart constructor.

Constructors

CancelAuditMitigationActionsTaskResponse'

Fields

httpStatus :: Int

The response's http status code.

newCancelAuditMitigationActionsTaskResponse

Arguments

:: Int
$sel:httpStatus:CancelAuditMitigationActionsTaskResponse'

-> CancelAuditMitigationActionsTaskResponse

Create a value of CancelAuditMitigationActionsTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CancelAuditMitigationActionsTaskResponse', cancelAuditMitigationActionsTaskResponse_httpStatus - The response's http status code.

Response Lenses

cancelAuditMitigationActionsTaskResponse_httpStatus :: Lens' CancelAuditMitigationActionsTaskResponse Int

The response's http status code.