Network.AWS.IoT.CancelAuditTask

Description

Cancels an audit that is in progress. The audit can be either scheduled or on demand. If the audit isn't in progress, an "InvalidRequestException" occurs.

Requires permission to access the CancelAuditTask action.

Creating a Request

data CancelAuditTask

See: newCancelAuditTask smart constructor.

Constructors

CancelAuditTask'

Fields

taskId :: Text

The ID of the audit you want to cancel. You can only cancel an audit that is "IN_PROGRESS".

newCancelAuditTask

Arguments

:: Text
$sel:taskId:CancelAuditTask'

-> CancelAuditTask

Create a value of CancelAuditTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskId:CancelAuditTask', cancelAuditTask_taskId - The ID of the audit you want to cancel. You can only cancel an audit that is "IN_PROGRESS".

Request Lenses

cancelAuditTask_taskId :: Lens' CancelAuditTask Text

The ID of the audit you want to cancel. You can only cancel an audit that is "IN_PROGRESS".

Destructuring the Response

data CancelAuditTaskResponse

See: newCancelAuditTaskResponse smart constructor.

Constructors

CancelAuditTaskResponse'

Fields

httpStatus :: Int

The response's http status code.

newCancelAuditTaskResponse

Arguments

:: Int
$sel:httpStatus:CancelAuditTaskResponse'

-> CancelAuditTaskResponse

Create a value of CancelAuditTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CancelAuditTaskResponse', cancelAuditTaskResponse_httpStatus - The response's http status code.

Response Lenses

cancelAuditTaskResponse_httpStatus :: Lens' CancelAuditTaskResponse Int

The response's http status code.