Network.AWS.IoT.CancelCertificateTransfer

Description

Cancels a pending transfer for the specified certificate.

Note Only the transfer source account can use this operation to cancel a transfer. (Transfer destinations can use RejectCertificateTransfer instead.) After transfer, IoT returns the certificate to the source account in the INACTIVE state. After the destination account has accepted the transfer, the transfer cannot be cancelled.

After a certificate transfer is cancelled, the status of the certificate changes from PENDING_TRANSFER to INACTIVE.

Requires permission to access the CancelCertificateTransfer action.

Creating a Request

data CancelCertificateTransfer

The input for the CancelCertificateTransfer operation.

See: newCancelCertificateTransfer smart constructor.

Constructors

CancelCertificateTransfer'

Fields

certificateId :: Text

The ID of the certificate. (The last part of the certificate ARN contains the certificate ID.)

newCancelCertificateTransfer

Arguments

:: Text
$sel:certificateId:CancelCertificateTransfer'

-> CancelCertificateTransfer

Create a value of CancelCertificateTransfer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateId:CancelCertificateTransfer', cancelCertificateTransfer_certificateId - The ID of the certificate. (The last part of the certificate ARN contains the certificate ID.)

Request Lenses

cancelCertificateTransfer_certificateId :: Lens' CancelCertificateTransfer Text

The ID of the certificate. (The last part of the certificate ARN contains the certificate ID.)

Destructuring the Response

data CancelCertificateTransferResponse

See: newCancelCertificateTransferResponse smart constructor.

Constructors

CancelCertificateTransferResponse'

newCancelCertificateTransferResponse :: CancelCertificateTransferResponse

Create a value of CancelCertificateTransferResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.