Network.AWS.IoT.ConfirmTopicRuleDestination

Description

Confirms a topic rule destination. When you create a rule requiring a destination, IoT sends a confirmation message to the endpoint or base address you specify. The message includes a token which you pass back when calling ConfirmTopicRuleDestination to confirm that you own or have access to the endpoint.

Requires permission to access the ConfirmTopicRuleDestination action.

Creating a Request

data ConfirmTopicRuleDestination

See: newConfirmTopicRuleDestination smart constructor.

Constructors

ConfirmTopicRuleDestination'

Fields

confirmationToken :: Text

The token used to confirm ownership or access to the topic rule confirmation URL.

newConfirmTopicRuleDestination

Arguments

:: Text
$sel:confirmationToken:ConfirmTopicRuleDestination'

-> ConfirmTopicRuleDestination

Create a value of ConfirmTopicRuleDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:confirmationToken:ConfirmTopicRuleDestination', confirmTopicRuleDestination_confirmationToken - The token used to confirm ownership or access to the topic rule confirmation URL.

Request Lenses

confirmTopicRuleDestination_confirmationToken :: Lens' ConfirmTopicRuleDestination Text

The token used to confirm ownership or access to the topic rule confirmation URL.

Destructuring the Response

data ConfirmTopicRuleDestinationResponse

See: newConfirmTopicRuleDestinationResponse smart constructor.

Constructors

ConfirmTopicRuleDestinationResponse'

Fields

httpStatus :: Int

The response's http status code.

newConfirmTopicRuleDestinationResponse

Arguments

:: Int
$sel:httpStatus:ConfirmTopicRuleDestinationResponse'

-> ConfirmTopicRuleDestinationResponse

Create a value of ConfirmTopicRuleDestinationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ConfirmTopicRuleDestinationResponse', confirmTopicRuleDestinationResponse_httpStatus - The response's http status code.

Response Lenses

confirmTopicRuleDestinationResponse_httpStatus :: Lens' ConfirmTopicRuleDestinationResponse Int

The response's http status code.