Network.AWS.IoT.CreateAuditSuppression

Description

Creates a Device Defender audit suppression.

Requires permission to access the CreateAuditSuppression action.

Creating a Request

data CreateAuditSuppression

See: newCreateAuditSuppression smart constructor.

Constructors

CreateAuditSuppression'

Fields

expirationDate :: Maybe POSIX

The epoch timestamp in seconds at which this suppression expires.

description :: Maybe Text

The description of the audit suppression.

suppressIndefinitely :: Maybe Bool

Indicates whether a suppression should exist indefinitely or not.

checkName :: Text

resourceIdentifier :: ResourceIdentifier

clientRequestToken :: Text

Each audit supression must have a unique client request token. If you try to create a new audit suppression with the same token as one that already exists, an exception occurs. If you omit this value, Amazon Web Services SDKs will automatically generate a unique client request.

newCreateAuditSuppression

Arguments

:: Text
$sel:checkName:CreateAuditSuppression'

-> ResourceIdentifier
$sel:resourceIdentifier:CreateAuditSuppression'

-> Text
$sel:clientRequestToken:CreateAuditSuppression'

-> CreateAuditSuppression

Create a value of CreateAuditSuppression with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expirationDate:CreateAuditSuppression', createAuditSuppression_expirationDate - The epoch timestamp in seconds at which this suppression expires.

$sel:description:CreateAuditSuppression', createAuditSuppression_description - The description of the audit suppression.

$sel:suppressIndefinitely:CreateAuditSuppression', createAuditSuppression_suppressIndefinitely - Indicates whether a suppression should exist indefinitely or not.

$sel:checkName:CreateAuditSuppression', createAuditSuppression_checkName - Undocumented member.

$sel:resourceIdentifier:CreateAuditSuppression', createAuditSuppression_resourceIdentifier - Undocumented member.

$sel:clientRequestToken:CreateAuditSuppression', createAuditSuppression_clientRequestToken - Each audit supression must have a unique client request token. If you try to create a new audit suppression with the same token as one that already exists, an exception occurs. If you omit this value, Amazon Web Services SDKs will automatically generate a unique client request.

Request Lenses

createAuditSuppression_expirationDate :: Lens' CreateAuditSuppression (Maybe UTCTime)

The epoch timestamp in seconds at which this suppression expires.

createAuditSuppression_description :: Lens' CreateAuditSuppression (Maybe Text)

The description of the audit suppression.

createAuditSuppression_suppressIndefinitely :: Lens' CreateAuditSuppression (Maybe Bool)

Indicates whether a suppression should exist indefinitely or not.

createAuditSuppression_checkName :: Lens' CreateAuditSuppression Text

Undocumented member.

createAuditSuppression_resourceIdentifier :: Lens' CreateAuditSuppression ResourceIdentifier

Undocumented member.

createAuditSuppression_clientRequestToken :: Lens' CreateAuditSuppression Text

Each audit supression must have a unique client request token. If you try to create a new audit suppression with the same token as one that already exists, an exception occurs. If you omit this value, Amazon Web Services SDKs will automatically generate a unique client request.

Destructuring the Response

data CreateAuditSuppressionResponse

See: newCreateAuditSuppressionResponse smart constructor.

Constructors

CreateAuditSuppressionResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateAuditSuppressionResponse

Arguments

:: Int
$sel:httpStatus:CreateAuditSuppressionResponse'

-> CreateAuditSuppressionResponse

Create a value of CreateAuditSuppressionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateAuditSuppressionResponse', createAuditSuppressionResponse_httpStatus - The response's http status code.

Response Lenses

createAuditSuppressionResponse_httpStatus :: Lens' CreateAuditSuppressionResponse Int

The response's http status code.