Network.AWS.IoT.CreateAuthorizer

Description

Creates an authorizer.

Requires permission to access the CreateAuthorizer action.

Creating a Request

data CreateAuthorizer

See: newCreateAuthorizer smart constructor.

Constructors

CreateAuthorizer'

Fields

status :: Maybe AuthorizerStatus

The status of the create authorizer request.

signingDisabled :: Maybe Bool

Specifies whether IoT validates the token signature in an authorization request.

tokenSigningPublicKeys :: Maybe (HashMap Text Text)

The public keys used to verify the digital signature returned by your custom authentication service.

tokenKeyName :: Maybe Text

The name of the token key used to extract the token from the HTTP headers.

tags :: Maybe [Tag]

Metadata which can be used to manage the custom authorizer.

For URI Request parameters use format: ...key1=value1&key2=value2...

For the CLI command-line parameter use format: &&tags "key1=value1&key2=value2..."

For the cli-input-json file use format: "tags": "key1=value1&key2=value2..."

authorizerName :: Text

The authorizer name.

authorizerFunctionArn :: Text

The ARN of the authorizer's Lambda function.

newCreateAuthorizer

Arguments

:: Text
$sel:authorizerName:CreateAuthorizer'

-> Text
$sel:authorizerFunctionArn:CreateAuthorizer'

-> CreateAuthorizer

Create a value of CreateAuthorizer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateAuthorizer', createAuthorizer_status - The status of the create authorizer request.

$sel:signingDisabled:CreateAuthorizer', createAuthorizer_signingDisabled - Specifies whether IoT validates the token signature in an authorization request.

$sel:tokenSigningPublicKeys:CreateAuthorizer', createAuthorizer_tokenSigningPublicKeys - The public keys used to verify the digital signature returned by your custom authentication service.

$sel:tokenKeyName:CreateAuthorizer', createAuthorizer_tokenKeyName - The name of the token key used to extract the token from the HTTP headers.

$sel:tags:CreateAuthorizer', createAuthorizer_tags - Metadata which can be used to manage the custom authorizer.

For URI Request parameters use format: ...key1=value1&key2=value2...

For the CLI command-line parameter use format: &&tags "key1=value1&key2=value2..."

For the cli-input-json file use format: "tags": "key1=value1&key2=value2..."

$sel:authorizerName:CreateAuthorizer', createAuthorizer_authorizerName - The authorizer name.

$sel:authorizerFunctionArn:CreateAuthorizer', createAuthorizer_authorizerFunctionArn - The ARN of the authorizer's Lambda function.

Request Lenses

createAuthorizer_status :: Lens' CreateAuthorizer (Maybe AuthorizerStatus)

The status of the create authorizer request.

createAuthorizer_signingDisabled :: Lens' CreateAuthorizer (Maybe Bool)

Specifies whether IoT validates the token signature in an authorization request.

createAuthorizer_tokenSigningPublicKeys :: Lens' CreateAuthorizer (Maybe (HashMap Text Text))

The public keys used to verify the digital signature returned by your custom authentication service.

createAuthorizer_tokenKeyName :: Lens' CreateAuthorizer (Maybe Text)

The name of the token key used to extract the token from the HTTP headers.

createAuthorizer_tags :: Lens' CreateAuthorizer (Maybe [Tag])

Metadata which can be used to manage the custom authorizer.

For URI Request parameters use format: ...key1=value1&key2=value2...

For the CLI command-line parameter use format: &&tags "key1=value1&key2=value2..."

For the cli-input-json file use format: "tags": "key1=value1&key2=value2..."

createAuthorizer_authorizerName :: Lens' CreateAuthorizer Text

The authorizer name.

createAuthorizer_authorizerFunctionArn :: Lens' CreateAuthorizer Text

The ARN of the authorizer's Lambda function.

Destructuring the Response

data CreateAuthorizerResponse

See: newCreateAuthorizerResponse smart constructor.

Constructors

CreateAuthorizerResponse'

Fields

authorizerName :: Maybe Text

The authorizer's name.

authorizerArn :: Maybe Text

The authorizer ARN.

httpStatus :: Int

The response's http status code.

newCreateAuthorizerResponse

Arguments

:: Int
$sel:httpStatus:CreateAuthorizerResponse'

-> CreateAuthorizerResponse

Create a value of CreateAuthorizerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authorizerName:CreateAuthorizer', createAuthorizerResponse_authorizerName - The authorizer's name.

$sel:authorizerArn:CreateAuthorizerResponse', createAuthorizerResponse_authorizerArn - The authorizer ARN.

$sel:httpStatus:CreateAuthorizerResponse', createAuthorizerResponse_httpStatus - The response's http status code.

Response Lenses

createAuthorizerResponse_authorizerName :: Lens' CreateAuthorizerResponse (Maybe Text)

The authorizer's name.

createAuthorizerResponse_authorizerArn :: Lens' CreateAuthorizerResponse (Maybe Text)

The authorizer ARN.

createAuthorizerResponse_httpStatus :: Lens' CreateAuthorizerResponse Int

The response's http status code.