Network.AWS.IoT.CreateBillingGroup

Description

Creates a billing group.

Requires permission to access the CreateBillingGroup action.

Creating a Request

data CreateBillingGroup

See: newCreateBillingGroup smart constructor.

Constructors

CreateBillingGroup'

Fields

billingGroupProperties :: Maybe BillingGroupProperties

The properties of the billing group.

tags :: Maybe [Tag]

Metadata which can be used to manage the billing group.

billingGroupName :: Text

The name you wish to give to the billing group.

newCreateBillingGroup

Arguments

:: Text
$sel:billingGroupName:CreateBillingGroup'

-> CreateBillingGroup

Create a value of CreateBillingGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:billingGroupProperties:CreateBillingGroup', createBillingGroup_billingGroupProperties - The properties of the billing group.

$sel:tags:CreateBillingGroup', createBillingGroup_tags - Metadata which can be used to manage the billing group.

$sel:billingGroupName:CreateBillingGroup', createBillingGroup_billingGroupName - The name you wish to give to the billing group.

Request Lenses

createBillingGroup_billingGroupProperties :: Lens' CreateBillingGroup (Maybe BillingGroupProperties)

The properties of the billing group.

createBillingGroup_tags :: Lens' CreateBillingGroup (Maybe [Tag])

Metadata which can be used to manage the billing group.

createBillingGroup_billingGroupName :: Lens' CreateBillingGroup Text

The name you wish to give to the billing group.

Destructuring the Response

data CreateBillingGroupResponse

See: newCreateBillingGroupResponse smart constructor.

Constructors

CreateBillingGroupResponse'

Fields

billingGroupArn :: Maybe Text

The ARN of the billing group.

billingGroupName :: Maybe Text

The name you gave to the billing group.

billingGroupId :: Maybe Text

The ID of the billing group.

httpStatus :: Int

The response's http status code.

newCreateBillingGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateBillingGroupResponse'

-> CreateBillingGroupResponse

Create a value of CreateBillingGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:billingGroupArn:CreateBillingGroupResponse', createBillingGroupResponse_billingGroupArn - The ARN of the billing group.

$sel:billingGroupName:CreateBillingGroup', createBillingGroupResponse_billingGroupName - The name you gave to the billing group.

$sel:billingGroupId:CreateBillingGroupResponse', createBillingGroupResponse_billingGroupId - The ID of the billing group.

$sel:httpStatus:CreateBillingGroupResponse', createBillingGroupResponse_httpStatus - The response's http status code.

Response Lenses

createBillingGroupResponse_billingGroupArn :: Lens' CreateBillingGroupResponse (Maybe Text)

The ARN of the billing group.

createBillingGroupResponse_billingGroupName :: Lens' CreateBillingGroupResponse (Maybe Text)

The name you gave to the billing group.

createBillingGroupResponse_billingGroupId :: Lens' CreateBillingGroupResponse (Maybe Text)

The ID of the billing group.

createBillingGroupResponse_httpStatus :: Lens' CreateBillingGroupResponse Int

The response's http status code.