Network.AWS.IoT.CreateCustomMetric

Description

Use this API to define a Custom Metric published by your devices to Device Defender.

Requires permission to access the CreateCustomMetric action.

Creating a Request

data CreateCustomMetric

See: newCreateCustomMetric smart constructor.

Constructors

CreateCustomMetric'

Fields

tags :: Maybe [Tag]

Metadata that can be used to manage the custom metric.

displayName :: Maybe Text

Field represents a friendly name in the console for the custom metric; it doesn't have to be unique. Don't use this name as the metric identifier in the device metric report. Can be updated once defined.

metricName :: Text

The name of the custom metric. This will be used in the metric report submitted from the device/thing. Shouldn't begin with aws:. Cannot be updated once defined.

metricType :: CustomMetricType

The type of the custom metric. Types include string-list, ip-address-list, number-list, and number.

clientRequestToken :: Text

Each custom metric must have a unique client request token. If you try to create a new custom metric that already exists with a different token, an exception occurs. If you omit this value, Amazon Web Services SDKs will automatically generate a unique client request.

newCreateCustomMetric

Arguments

:: Text
$sel:metricName:CreateCustomMetric'

-> CustomMetricType
$sel:metricType:CreateCustomMetric'

-> Text
$sel:clientRequestToken:CreateCustomMetric'

-> CreateCustomMetric

Create a value of CreateCustomMetric with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateCustomMetric', createCustomMetric_tags - Metadata that can be used to manage the custom metric.

$sel:displayName:CreateCustomMetric', createCustomMetric_displayName - Field represents a friendly name in the console for the custom metric; it doesn't have to be unique. Don't use this name as the metric identifier in the device metric report. Can be updated once defined.

$sel:metricName:CreateCustomMetric', createCustomMetric_metricName - The name of the custom metric. This will be used in the metric report submitted from the device/thing. Shouldn't begin with aws:. Cannot be updated once defined.

$sel:metricType:CreateCustomMetric', createCustomMetric_metricType - The type of the custom metric. Types include string-list, ip-address-list, number-list, and number.

$sel:clientRequestToken:CreateCustomMetric', createCustomMetric_clientRequestToken - Each custom metric must have a unique client request token. If you try to create a new custom metric that already exists with a different token, an exception occurs. If you omit this value, Amazon Web Services SDKs will automatically generate a unique client request.

Request Lenses

createCustomMetric_tags :: Lens' CreateCustomMetric (Maybe [Tag])

Metadata that can be used to manage the custom metric.

createCustomMetric_displayName :: Lens' CreateCustomMetric (Maybe Text)

Field represents a friendly name in the console for the custom metric; it doesn't have to be unique. Don't use this name as the metric identifier in the device metric report. Can be updated once defined.

createCustomMetric_metricName :: Lens' CreateCustomMetric Text

The name of the custom metric. This will be used in the metric report submitted from the device/thing. Shouldn't begin with aws:. Cannot be updated once defined.

createCustomMetric_metricType :: Lens' CreateCustomMetric CustomMetricType

The type of the custom metric. Types include string-list, ip-address-list, number-list, and number.

createCustomMetric_clientRequestToken :: Lens' CreateCustomMetric Text

Each custom metric must have a unique client request token. If you try to create a new custom metric that already exists with a different token, an exception occurs. If you omit this value, Amazon Web Services SDKs will automatically generate a unique client request.

Destructuring the Response

data CreateCustomMetricResponse

See: newCreateCustomMetricResponse smart constructor.

Constructors

CreateCustomMetricResponse'

Fields

metricArn :: Maybe Text

The Amazon Resource Number (ARN) of the custom metric, e.g. arn:aws-partition:iot:region:accountId:custommetric/metricName

metricName :: Maybe Text

The name of the custom metric to be used in the metric report.

httpStatus :: Int

The response's http status code.

newCreateCustomMetricResponse

Arguments

:: Int
$sel:httpStatus:CreateCustomMetricResponse'

-> CreateCustomMetricResponse

Create a value of CreateCustomMetricResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricArn:CreateCustomMetricResponse', createCustomMetricResponse_metricArn - The Amazon Resource Number (ARN) of the custom metric, e.g. arn:aws-partition:iot:region:accountId:custommetric/metricName

$sel:metricName:CreateCustomMetric', createCustomMetricResponse_metricName - The name of the custom metric to be used in the metric report.

$sel:httpStatus:CreateCustomMetricResponse', createCustomMetricResponse_httpStatus - The response's http status code.

Response Lenses

createCustomMetricResponse_metricArn :: Lens' CreateCustomMetricResponse (Maybe Text)

The Amazon Resource Number (ARN) of the custom metric, e.g. arn:aws-partition:iot:region:accountId:custommetric/metricName

createCustomMetricResponse_metricName :: Lens' CreateCustomMetricResponse (Maybe Text)

The name of the custom metric to be used in the metric report.

createCustomMetricResponse_httpStatus :: Lens' CreateCustomMetricResponse Int

The response's http status code.