Network.AWS.IoT.CreateDimension

Description

Create a dimension that you can use to limit the scope of a metric used in a security profile for IoT Device Defender. For example, using a TOPIC_FILTER dimension, you can narrow down the scope of the metric only to MQTT topics whose name match the pattern specified in the dimension.

Requires permission to access the CreateDimension action.

Creating a Request

data CreateDimension

See: newCreateDimension smart constructor.

Constructors

CreateDimension'

Fields

tags :: Maybe [Tag]

Metadata that can be used to manage the dimension.

name :: Text

A unique identifier for the dimension. Choose something that describes the type and value to make it easy to remember what it does.

type' :: DimensionType

Specifies the type of dimension. Supported types: TOPIC_FILTER.

stringValues :: NonEmpty Text

Specifies the value or list of values for the dimension. For TOPIC_FILTER dimensions, this is a pattern used to match the MQTT topic (for example, "admin/#").

clientRequestToken :: Text

Each dimension must have a unique client request token. If you try to create a new dimension with the same token as a dimension that already exists, an exception occurs. If you omit this value, Amazon Web Services SDKs will automatically generate a unique client request.

newCreateDimension

Arguments

:: Text
$sel:name:CreateDimension'

-> DimensionType
$sel:type':CreateDimension'

-> NonEmpty Text
$sel:stringValues:CreateDimension'

-> Text
$sel:clientRequestToken:CreateDimension'

-> CreateDimension

Create a value of CreateDimension with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateDimension', createDimension_tags - Metadata that can be used to manage the dimension.

$sel:name:CreateDimension', createDimension_name - A unique identifier for the dimension. Choose something that describes the type and value to make it easy to remember what it does.

$sel:type':CreateDimension', createDimension_type - Specifies the type of dimension. Supported types: TOPIC_FILTER.

$sel:stringValues:CreateDimension', createDimension_stringValues - Specifies the value or list of values for the dimension. For TOPIC_FILTER dimensions, this is a pattern used to match the MQTT topic (for example, "admin/#").

$sel:clientRequestToken:CreateDimension', createDimension_clientRequestToken - Each dimension must have a unique client request token. If you try to create a new dimension with the same token as a dimension that already exists, an exception occurs. If you omit this value, Amazon Web Services SDKs will automatically generate a unique client request.

Request Lenses

createDimension_tags :: Lens' CreateDimension (Maybe [Tag])

Metadata that can be used to manage the dimension.

createDimension_name :: Lens' CreateDimension Text

A unique identifier for the dimension. Choose something that describes the type and value to make it easy to remember what it does.

createDimension_type :: Lens' CreateDimension DimensionType

Specifies the type of dimension. Supported types: TOPIC_FILTER.

createDimension_stringValues :: Lens' CreateDimension (NonEmpty Text)

Specifies the value or list of values for the dimension. For TOPIC_FILTER dimensions, this is a pattern used to match the MQTT topic (for example, "admin/#").

createDimension_clientRequestToken :: Lens' CreateDimension Text

Each dimension must have a unique client request token. If you try to create a new dimension with the same token as a dimension that already exists, an exception occurs. If you omit this value, Amazon Web Services SDKs will automatically generate a unique client request.

Destructuring the Response

data CreateDimensionResponse

See: newCreateDimensionResponse smart constructor.

Constructors

CreateDimensionResponse'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the created dimension.

name :: Maybe Text

A unique identifier for the dimension.

httpStatus :: Int

The response's http status code.

newCreateDimensionResponse

Arguments

:: Int
$sel:httpStatus:CreateDimensionResponse'

-> CreateDimensionResponse

Create a value of CreateDimensionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:CreateDimensionResponse', createDimensionResponse_arn - The Amazon Resource Name (ARN) of the created dimension.

$sel:name:CreateDimension', createDimensionResponse_name - A unique identifier for the dimension.

$sel:httpStatus:CreateDimensionResponse', createDimensionResponse_httpStatus - The response's http status code.

Response Lenses

createDimensionResponse_arn :: Lens' CreateDimensionResponse (Maybe Text)

The Amazon Resource Name (ARN) of the created dimension.

createDimensionResponse_name :: Lens' CreateDimensionResponse (Maybe Text)

A unique identifier for the dimension.

createDimensionResponse_httpStatus :: Lens' CreateDimensionResponse Int

The response's http status code.