Network.AWS.IoT.CreateDomainConfiguration

Description

Creates a domain configuration.

Requires permission to access the CreateDomainConfiguration action.

Creating a Request

data CreateDomainConfiguration

See: newCreateDomainConfiguration smart constructor.

Constructors

CreateDomainConfiguration'

Fields

authorizerConfig :: Maybe AuthorizerConfig

An object that specifies the authorization service for a domain.

serverCertificateArns :: Maybe [Text]

The ARNs of the certificates that IoT passes to the device during the TLS handshake. Currently you can specify only one certificate ARN. This value is not required for Amazon Web Services-managed domains.

domainName :: Maybe Text

The name of the domain.

serviceType :: Maybe ServiceType

The type of service delivered by the endpoint.

Amazon Web Services IoT Core currently supports only the DATA service type.

validationCertificateArn :: Maybe Text

The certificate used to validate the server certificate and prove domain name ownership. This certificate must be signed by a public certificate authority. This value is not required for Amazon Web Services-managed domains.

tags :: Maybe [Tag]

Metadata which can be used to manage the domain configuration.

For URI Request parameters use format: ...key1=value1&key2=value2...

For the CLI command-line parameter use format: &&tags "key1=value1&key2=value2..."

For the cli-input-json file use format: "tags": "key1=value1&key2=value2..."

domainConfigurationName :: Text

The name of the domain configuration. This value must be unique to a region.

newCreateDomainConfiguration

Arguments

:: Text
$sel:domainConfigurationName:CreateDomainConfiguration'

-> CreateDomainConfiguration

Create a value of CreateDomainConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authorizerConfig:CreateDomainConfiguration', createDomainConfiguration_authorizerConfig - An object that specifies the authorization service for a domain.

$sel:serverCertificateArns:CreateDomainConfiguration', createDomainConfiguration_serverCertificateArns - The ARNs of the certificates that IoT passes to the device during the TLS handshake. Currently you can specify only one certificate ARN. This value is not required for Amazon Web Services-managed domains.

$sel:domainName:CreateDomainConfiguration', createDomainConfiguration_domainName - The name of the domain.

$sel:serviceType:CreateDomainConfiguration', createDomainConfiguration_serviceType - The type of service delivered by the endpoint.

Amazon Web Services IoT Core currently supports only the DATA service type.

$sel:validationCertificateArn:CreateDomainConfiguration', createDomainConfiguration_validationCertificateArn - The certificate used to validate the server certificate and prove domain name ownership. This certificate must be signed by a public certificate authority. This value is not required for Amazon Web Services-managed domains.

$sel:tags:CreateDomainConfiguration', createDomainConfiguration_tags - Metadata which can be used to manage the domain configuration.

For URI Request parameters use format: ...key1=value1&key2=value2...

For the CLI command-line parameter use format: &&tags "key1=value1&key2=value2..."

For the cli-input-json file use format: "tags": "key1=value1&key2=value2..."

$sel:domainConfigurationName:CreateDomainConfiguration', createDomainConfiguration_domainConfigurationName - The name of the domain configuration. This value must be unique to a region.

Request Lenses

createDomainConfiguration_authorizerConfig :: Lens' CreateDomainConfiguration (Maybe AuthorizerConfig)

An object that specifies the authorization service for a domain.

createDomainConfiguration_serverCertificateArns :: Lens' CreateDomainConfiguration (Maybe [Text])

The ARNs of the certificates that IoT passes to the device during the TLS handshake. Currently you can specify only one certificate ARN. This value is not required for Amazon Web Services-managed domains.

createDomainConfiguration_domainName :: Lens' CreateDomainConfiguration (Maybe Text)

The name of the domain.

createDomainConfiguration_serviceType :: Lens' CreateDomainConfiguration (Maybe ServiceType)

The type of service delivered by the endpoint.

Amazon Web Services IoT Core currently supports only the DATA service type.

createDomainConfiguration_validationCertificateArn :: Lens' CreateDomainConfiguration (Maybe Text)

The certificate used to validate the server certificate and prove domain name ownership. This certificate must be signed by a public certificate authority. This value is not required for Amazon Web Services-managed domains.

createDomainConfiguration_tags :: Lens' CreateDomainConfiguration (Maybe [Tag])

Metadata which can be used to manage the domain configuration.

For URI Request parameters use format: ...key1=value1&key2=value2...

For the CLI command-line parameter use format: &&tags "key1=value1&key2=value2..."

For the cli-input-json file use format: "tags": "key1=value1&key2=value2..."

createDomainConfiguration_domainConfigurationName :: Lens' CreateDomainConfiguration Text

The name of the domain configuration. This value must be unique to a region.

Destructuring the Response

data CreateDomainConfigurationResponse

See: newCreateDomainConfigurationResponse smart constructor.

Constructors

CreateDomainConfigurationResponse'

Fields

domainConfigurationName :: Maybe Text

The name of the domain configuration.

domainConfigurationArn :: Maybe Text

The ARN of the domain configuration.

httpStatus :: Int

The response's http status code.

newCreateDomainConfigurationResponse

Arguments

:: Int
$sel:httpStatus:CreateDomainConfigurationResponse'

-> CreateDomainConfigurationResponse

Create a value of CreateDomainConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainConfigurationName:CreateDomainConfiguration', createDomainConfigurationResponse_domainConfigurationName - The name of the domain configuration.

$sel:domainConfigurationArn:CreateDomainConfigurationResponse', createDomainConfigurationResponse_domainConfigurationArn - The ARN of the domain configuration.

$sel:httpStatus:CreateDomainConfigurationResponse', createDomainConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

createDomainConfigurationResponse_domainConfigurationName :: Lens' CreateDomainConfigurationResponse (Maybe Text)

The name of the domain configuration.

createDomainConfigurationResponse_domainConfigurationArn :: Lens' CreateDomainConfigurationResponse (Maybe Text)

The ARN of the domain configuration.

createDomainConfigurationResponse_httpStatus :: Lens' CreateDomainConfigurationResponse Int

The response's http status code.