Network.AWS.IoT.CreateDynamicThingGroup

Description

Creates a dynamic thing group.

Requires permission to access the CreateDynamicThingGroup action.

Creating a Request

data CreateDynamicThingGroup

See: newCreateDynamicThingGroup smart constructor.

Constructors

CreateDynamicThingGroup'

Fields

indexName :: Maybe Text

The dynamic thing group index name.

Currently one index is supported: AWS_Things.

queryVersion :: Maybe Text

The dynamic thing group query version.

Currently one query version is supported: "2017-09-30". If not specified, the query version defaults to this value.

tags :: Maybe [Tag]

Metadata which can be used to manage the dynamic thing group.

thingGroupProperties :: Maybe ThingGroupProperties

The dynamic thing group properties.

thingGroupName :: Text

The dynamic thing group name to create.

queryString :: Text

The dynamic thing group search query string.

See Query Syntax for information about query string syntax.

newCreateDynamicThingGroup

Arguments

:: Text
$sel:thingGroupName:CreateDynamicThingGroup'

-> Text
$sel:queryString:CreateDynamicThingGroup'

-> CreateDynamicThingGroup

Create a value of CreateDynamicThingGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:indexName:CreateDynamicThingGroup', createDynamicThingGroup_indexName - The dynamic thing group index name.

Currently one index is supported: AWS_Things.

$sel:queryVersion:CreateDynamicThingGroup', createDynamicThingGroup_queryVersion - The dynamic thing group query version.

Currently one query version is supported: "2017-09-30". If not specified, the query version defaults to this value.

$sel:tags:CreateDynamicThingGroup', createDynamicThingGroup_tags - Metadata which can be used to manage the dynamic thing group.

$sel:thingGroupProperties:CreateDynamicThingGroup', createDynamicThingGroup_thingGroupProperties - The dynamic thing group properties.

$sel:thingGroupName:CreateDynamicThingGroup', createDynamicThingGroup_thingGroupName - The dynamic thing group name to create.

$sel:queryString:CreateDynamicThingGroup', createDynamicThingGroup_queryString - The dynamic thing group search query string.

See Query Syntax for information about query string syntax.

Request Lenses

createDynamicThingGroup_indexName :: Lens' CreateDynamicThingGroup (Maybe Text)

The dynamic thing group index name.

Currently one index is supported: AWS_Things.

createDynamicThingGroup_queryVersion :: Lens' CreateDynamicThingGroup (Maybe Text)

The dynamic thing group query version.

Currently one query version is supported: "2017-09-30". If not specified, the query version defaults to this value.

createDynamicThingGroup_tags :: Lens' CreateDynamicThingGroup (Maybe [Tag])

Metadata which can be used to manage the dynamic thing group.

createDynamicThingGroup_thingGroupProperties :: Lens' CreateDynamicThingGroup (Maybe ThingGroupProperties)

The dynamic thing group properties.

createDynamicThingGroup_thingGroupName :: Lens' CreateDynamicThingGroup Text

The dynamic thing group name to create.

createDynamicThingGroup_queryString :: Lens' CreateDynamicThingGroup Text

The dynamic thing group search query string.

See Query Syntax for information about query string syntax.

Destructuring the Response

data CreateDynamicThingGroupResponse

See: newCreateDynamicThingGroupResponse smart constructor.

Constructors

CreateDynamicThingGroupResponse'

Fields

queryString :: Maybe Text

The dynamic thing group search query string.

indexName :: Maybe Text

The dynamic thing group index name.

thingGroupArn :: Maybe Text

The dynamic thing group ARN.

queryVersion :: Maybe Text

The dynamic thing group query version.

thingGroupName :: Maybe Text

The dynamic thing group name.

thingGroupId :: Maybe Text

The dynamic thing group ID.

httpStatus :: Int

The response's http status code.

newCreateDynamicThingGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateDynamicThingGroupResponse'

-> CreateDynamicThingGroupResponse

Create a value of CreateDynamicThingGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queryString:CreateDynamicThingGroup', createDynamicThingGroupResponse_queryString - The dynamic thing group search query string.

$sel:indexName:CreateDynamicThingGroup', createDynamicThingGroupResponse_indexName - The dynamic thing group index name.

$sel:thingGroupArn:CreateDynamicThingGroupResponse', createDynamicThingGroupResponse_thingGroupArn - The dynamic thing group ARN.

$sel:queryVersion:CreateDynamicThingGroup', createDynamicThingGroupResponse_queryVersion - The dynamic thing group query version.

$sel:thingGroupName:CreateDynamicThingGroup', createDynamicThingGroupResponse_thingGroupName - The dynamic thing group name.

$sel:thingGroupId:CreateDynamicThingGroupResponse', createDynamicThingGroupResponse_thingGroupId - The dynamic thing group ID.

$sel:httpStatus:CreateDynamicThingGroupResponse', createDynamicThingGroupResponse_httpStatus - The response's http status code.

Response Lenses

createDynamicThingGroupResponse_queryString :: Lens' CreateDynamicThingGroupResponse (Maybe Text)

The dynamic thing group search query string.

createDynamicThingGroupResponse_indexName :: Lens' CreateDynamicThingGroupResponse (Maybe Text)

The dynamic thing group index name.

createDynamicThingGroupResponse_thingGroupArn :: Lens' CreateDynamicThingGroupResponse (Maybe Text)

The dynamic thing group ARN.

createDynamicThingGroupResponse_queryVersion :: Lens' CreateDynamicThingGroupResponse (Maybe Text)

The dynamic thing group query version.

createDynamicThingGroupResponse_thingGroupName :: Lens' CreateDynamicThingGroupResponse (Maybe Text)

The dynamic thing group name.

createDynamicThingGroupResponse_thingGroupId :: Lens' CreateDynamicThingGroupResponse (Maybe Text)

The dynamic thing group ID.

createDynamicThingGroupResponse_httpStatus :: Lens' CreateDynamicThingGroupResponse Int

The response's http status code.