Network.AWS.IoT.CreateFleetMetric

Description

Creates a fleet metric.

Requires permission to access the CreateFleetMetric action.

Creating a Request

data CreateFleetMetric

See: newCreateFleetMetric smart constructor.

Constructors

CreateFleetMetric'

Fields

indexName :: Maybe Text

The name of the index to search.

unit :: Maybe FleetMetricUnit

Used to support unit transformation such as milliseconds to seconds. The unit must be supported by CW metric. Default to null.

queryVersion :: Maybe Text

The query version.

tags :: Maybe [Tag]

Metadata, which can be used to manage the fleet metric.

description :: Maybe Text

The fleet metric description.

metricName :: Text

The name of the fleet metric to create.

queryString :: Text

The search query string.

aggregationType :: AggregationType

The type of the aggregation query.

period :: Natural

The time in seconds between fleet metric emissions. Range [60(1 min), 86400(1 day)] and must be multiple of 60.

aggregationField :: Text

The field to aggregate.

newCreateFleetMetric

Arguments

:: Text
$sel:metricName:CreateFleetMetric'

-> Text
$sel:queryString:CreateFleetMetric'

-> AggregationType
$sel:aggregationType:CreateFleetMetric'

-> Natural
$sel:period:CreateFleetMetric'

-> Text
$sel:aggregationField:CreateFleetMetric'

-> CreateFleetMetric

Create a value of CreateFleetMetric with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:indexName:CreateFleetMetric', createFleetMetric_indexName - The name of the index to search.

$sel:unit:CreateFleetMetric', createFleetMetric_unit - Used to support unit transformation such as milliseconds to seconds. The unit must be supported by CW metric. Default to null.

$sel:queryVersion:CreateFleetMetric', createFleetMetric_queryVersion - The query version.

$sel:tags:CreateFleetMetric', createFleetMetric_tags - Metadata, which can be used to manage the fleet metric.

$sel:description:CreateFleetMetric', createFleetMetric_description - The fleet metric description.

$sel:metricName:CreateFleetMetric', createFleetMetric_metricName - The name of the fleet metric to create.

$sel:queryString:CreateFleetMetric', createFleetMetric_queryString - The search query string.

$sel:aggregationType:CreateFleetMetric', createFleetMetric_aggregationType - The type of the aggregation query.

$sel:period:CreateFleetMetric', createFleetMetric_period - The time in seconds between fleet metric emissions. Range [60(1 min), 86400(1 day)] and must be multiple of 60.

$sel:aggregationField:CreateFleetMetric', createFleetMetric_aggregationField - The field to aggregate.

Request Lenses

createFleetMetric_indexName :: Lens' CreateFleetMetric (Maybe Text)

The name of the index to search.

createFleetMetric_unit :: Lens' CreateFleetMetric (Maybe FleetMetricUnit)

Used to support unit transformation such as milliseconds to seconds. The unit must be supported by CW metric. Default to null.

createFleetMetric_queryVersion :: Lens' CreateFleetMetric (Maybe Text)

The query version.

createFleetMetric_tags :: Lens' CreateFleetMetric (Maybe [Tag])

Metadata, which can be used to manage the fleet metric.

createFleetMetric_description :: Lens' CreateFleetMetric (Maybe Text)

The fleet metric description.

createFleetMetric_metricName :: Lens' CreateFleetMetric Text

The name of the fleet metric to create.

createFleetMetric_queryString :: Lens' CreateFleetMetric Text

The search query string.

createFleetMetric_aggregationType :: Lens' CreateFleetMetric AggregationType

The type of the aggregation query.

createFleetMetric_period :: Lens' CreateFleetMetric Natural

The time in seconds between fleet metric emissions. Range [60(1 min), 86400(1 day)] and must be multiple of 60.

createFleetMetric_aggregationField :: Lens' CreateFleetMetric Text

The field to aggregate.

Destructuring the Response

data CreateFleetMetricResponse

See: newCreateFleetMetricResponse smart constructor.

Constructors

CreateFleetMetricResponse'

Fields

metricArn :: Maybe Text

The Amazon Resource Name (ARN) of the new fleet metric.

metricName :: Maybe Text

The name of the fleet metric to create.

httpStatus :: Int

The response's http status code.

newCreateFleetMetricResponse

Arguments

:: Int
$sel:httpStatus:CreateFleetMetricResponse'

-> CreateFleetMetricResponse

Create a value of CreateFleetMetricResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricArn:CreateFleetMetricResponse', createFleetMetricResponse_metricArn - The Amazon Resource Name (ARN) of the new fleet metric.

$sel:metricName:CreateFleetMetric', createFleetMetricResponse_metricName - The name of the fleet metric to create.

$sel:httpStatus:CreateFleetMetricResponse', createFleetMetricResponse_httpStatus - The response's http status code.

Response Lenses

createFleetMetricResponse_metricArn :: Lens' CreateFleetMetricResponse (Maybe Text)

The Amazon Resource Name (ARN) of the new fleet metric.

createFleetMetricResponse_metricName :: Lens' CreateFleetMetricResponse (Maybe Text)

The name of the fleet metric to create.

createFleetMetricResponse_httpStatus :: Lens' CreateFleetMetricResponse Int

The response's http status code.