Network.AWS.IoT.CreateJob

Description

Creates a job.

Requires permission to access the CreateJob action.

Creating a Request

data CreateJob

See: newCreateJob smart constructor.

Constructors

CreateJob'

Fields

jobExecutionsRolloutConfig :: Maybe JobExecutionsRolloutConfig

Allows you to create a staged rollout of the job.

documentSource :: Maybe Text

An S3 link to the job document. Required if you don't specify a value for document.

If the job document resides in an S3 bucket, you must use a placeholder link when specifying the document.

The placeholder link is of the following form:

${aws:iot:s3-presigned-url:https://s3.amazonaws.com/bucket/key}

where bucket is your bucket name and key is the object in the bucket to which you are linking.

abortConfig :: Maybe AbortConfig

Allows you to create criteria to abort a job.

namespaceId :: Maybe Text

The namespace used to indicate that a job is a customer-managed job.

When you specify a value for this parameter, Amazon Web Services IoT Core sends jobs notifications to MQTT topics that contain the value in the following format.

$aws/things/THING_NAME/jobs/JOB_ID/notify-namespace-NAMESPACE_ID/

The namespaceId feature is in public preview.

presignedUrlConfig :: Maybe PresignedUrlConfig

Configuration information for pre-signed S3 URLs.

document :: Maybe Text

The job document. Required if you don't specify a value for documentSource.

jobTemplateArn :: Maybe Text

The ARN of the job template used to create the job.

description :: Maybe Text

A short text description of the job.

targetSelection :: Maybe TargetSelection

Specifies whether the job will continue to run (CONTINUOUS), or will be complete after all those things specified as targets have completed the job (SNAPSHOT). If continuous, the job may also be run on a thing when a change is detected in a target. For example, a job will run on a thing when the thing is added to a target group, even after the job was completed by all things originally in the group.

timeoutConfig :: Maybe TimeoutConfig

Specifies the amount of time each device has to finish its execution of the job. The timer is started when the job execution status is set to IN_PROGRESS. If the job execution status is not set to another terminal state before the time expires, it will be automatically set to TIMED_OUT.

tags :: Maybe [Tag]

Metadata which can be used to manage the job.

jobId :: Text

A job identifier which must be unique for your Amazon Web Services account. We recommend using a UUID. Alpha-numeric characters, "-" and "_" are valid for use here.

targets :: NonEmpty Text

A list of things and thing groups to which the job should be sent.

newCreateJob

Arguments

:: Text
$sel:jobId:CreateJob'

-> NonEmpty Text
$sel:targets:CreateJob'

-> CreateJob

Create a value of CreateJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobExecutionsRolloutConfig:CreateJob', createJob_jobExecutionsRolloutConfig - Allows you to create a staged rollout of the job.

$sel:documentSource:CreateJob', createJob_documentSource - An S3 link to the job document. Required if you don't specify a value for document.

If the job document resides in an S3 bucket, you must use a placeholder link when specifying the document.

The placeholder link is of the following form:

${aws:iot:s3-presigned-url:https://s3.amazonaws.com/bucket/key}

where bucket is your bucket name and key is the object in the bucket to which you are linking.

$sel:abortConfig:CreateJob', createJob_abortConfig - Allows you to create criteria to abort a job.

$sel:namespaceId:CreateJob', createJob_namespaceId - The namespace used to indicate that a job is a customer-managed job.

When you specify a value for this parameter, Amazon Web Services IoT Core sends jobs notifications to MQTT topics that contain the value in the following format.

$aws/things/THING_NAME/jobs/JOB_ID/notify-namespace-NAMESPACE_ID/

The namespaceId feature is in public preview.

$sel:presignedUrlConfig:CreateJob', createJob_presignedUrlConfig - Configuration information for pre-signed S3 URLs.

$sel:document:CreateJob', createJob_document - The job document. Required if you don't specify a value for documentSource.

$sel:jobTemplateArn:CreateJob', createJob_jobTemplateArn - The ARN of the job template used to create the job.

$sel:description:CreateJob', createJob_description - A short text description of the job.

$sel:targetSelection:CreateJob', createJob_targetSelection - Specifies whether the job will continue to run (CONTINUOUS), or will be complete after all those things specified as targets have completed the job (SNAPSHOT). If continuous, the job may also be run on a thing when a change is detected in a target. For example, a job will run on a thing when the thing is added to a target group, even after the job was completed by all things originally in the group.

$sel:timeoutConfig:CreateJob', createJob_timeoutConfig - Specifies the amount of time each device has to finish its execution of the job. The timer is started when the job execution status is set to IN_PROGRESS. If the job execution status is not set to another terminal state before the time expires, it will be automatically set to TIMED_OUT.

$sel:tags:CreateJob', createJob_tags - Metadata which can be used to manage the job.

$sel:jobId:CreateJob', createJob_jobId - A job identifier which must be unique for your Amazon Web Services account. We recommend using a UUID. Alpha-numeric characters, "-" and "_" are valid for use here.

$sel:targets:CreateJob', createJob_targets - A list of things and thing groups to which the job should be sent.

Request Lenses

createJob_jobExecutionsRolloutConfig :: Lens' CreateJob (Maybe JobExecutionsRolloutConfig)

Allows you to create a staged rollout of the job.

createJob_documentSource :: Lens' CreateJob (Maybe Text)

An S3 link to the job document. Required if you don't specify a value for document.

If the job document resides in an S3 bucket, you must use a placeholder link when specifying the document.

The placeholder link is of the following form:

${aws:iot:s3-presigned-url:https://s3.amazonaws.com/bucket/key}

where bucket is your bucket name and key is the object in the bucket to which you are linking.

createJob_abortConfig :: Lens' CreateJob (Maybe AbortConfig)

Allows you to create criteria to abort a job.

createJob_namespaceId :: Lens' CreateJob (Maybe Text)

The namespace used to indicate that a job is a customer-managed job.

When you specify a value for this parameter, Amazon Web Services IoT Core sends jobs notifications to MQTT topics that contain the value in the following format.

$aws/things/THING_NAME/jobs/JOB_ID/notify-namespace-NAMESPACE_ID/

The namespaceId feature is in public preview.

createJob_presignedUrlConfig :: Lens' CreateJob (Maybe PresignedUrlConfig)

Configuration information for pre-signed S3 URLs.

createJob_document :: Lens' CreateJob (Maybe Text)

The job document. Required if you don't specify a value for documentSource.

createJob_jobTemplateArn :: Lens' CreateJob (Maybe Text)

The ARN of the job template used to create the job.

createJob_description :: Lens' CreateJob (Maybe Text)

A short text description of the job.

createJob_targetSelection :: Lens' CreateJob (Maybe TargetSelection)

Specifies whether the job will continue to run (CONTINUOUS), or will be complete after all those things specified as targets have completed the job (SNAPSHOT). If continuous, the job may also be run on a thing when a change is detected in a target. For example, a job will run on a thing when the thing is added to a target group, even after the job was completed by all things originally in the group.

createJob_timeoutConfig :: Lens' CreateJob (Maybe TimeoutConfig)

Specifies the amount of time each device has to finish its execution of the job. The timer is started when the job execution status is set to IN_PROGRESS. If the job execution status is not set to another terminal state before the time expires, it will be automatically set to TIMED_OUT.

createJob_tags :: Lens' CreateJob (Maybe [Tag])

Metadata which can be used to manage the job.

createJob_jobId :: Lens' CreateJob Text

A job identifier which must be unique for your Amazon Web Services account. We recommend using a UUID. Alpha-numeric characters, "-" and "_" are valid for use here.

createJob_targets :: Lens' CreateJob (NonEmpty Text)

A list of things and thing groups to which the job should be sent.

Destructuring the Response

data CreateJobResponse

See: newCreateJobResponse smart constructor.

Constructors

CreateJobResponse'

Fields

jobId :: Maybe Text

The unique identifier you assigned to this job.

jobArn :: Maybe Text

The job ARN.

description :: Maybe Text

The job description.

httpStatus :: Int

The response's http status code.

newCreateJobResponse

Arguments

:: Int
$sel:httpStatus:CreateJobResponse'

-> CreateJobResponse

Create a value of CreateJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:CreateJob', createJobResponse_jobId - The unique identifier you assigned to this job.

$sel:jobArn:CreateJobResponse', createJobResponse_jobArn - The job ARN.

$sel:description:CreateJob', createJobResponse_description - The job description.

$sel:httpStatus:CreateJobResponse', createJobResponse_httpStatus - The response's http status code.

Response Lenses

createJobResponse_jobId :: Lens' CreateJobResponse (Maybe Text)

The unique identifier you assigned to this job.

createJobResponse_jobArn :: Lens' CreateJobResponse (Maybe Text)

The job ARN.

createJobResponse_description :: Lens' CreateJobResponse (Maybe Text)

The job description.

createJobResponse_httpStatus :: Lens' CreateJobResponse Int

The response's http status code.