Network.AWS.IoT.CreateOTAUpdate

Description

Creates an IoT OTA update on a target group of things or groups.

Requires permission to access the CreateOTAUpdate action.

Creating a Request

data CreateOTAUpdate

See: newCreateOTAUpdate smart constructor.

Constructors

CreateOTAUpdate'

Fields

awsJobAbortConfig :: Maybe AwsJobAbortConfig

The criteria that determine when and how a job abort takes place.

awsJobExecutionsRolloutConfig :: Maybe AwsJobExecutionsRolloutConfig

Configuration for the rollout of OTA updates.

protocols :: Maybe (NonEmpty Protocol)

The protocol used to transfer the OTA update image. Valid values are [HTTP], [MQTT], [HTTP, MQTT]. When both HTTP and MQTT are specified, the target device can choose the protocol.

awsJobPresignedUrlConfig :: Maybe AwsJobPresignedUrlConfig

Configuration information for pre-signed URLs.

additionalParameters :: Maybe (HashMap Text Text)

A list of additional OTA update parameters which are name-value pairs.

awsJobTimeoutConfig :: Maybe AwsJobTimeoutConfig

Specifies the amount of time each device has to finish its execution of the job. A timer is started when the job execution status is set to IN_PROGRESS. If the job execution status is not set to another terminal state before the timer expires, it will be automatically set to TIMED_OUT.

description :: Maybe Text

The description of the OTA update.

targetSelection :: Maybe TargetSelection

Specifies whether the update will continue to run (CONTINUOUS), or will be complete after all the things specified as targets have completed the update (SNAPSHOT). If continuous, the update may also be run on a thing when a change is detected in a target. For example, an update will run on a thing when the thing is added to a target group, even after the update was completed by all things originally in the group. Valid values: CONTINUOUS | SNAPSHOT.

tags :: Maybe [Tag]

Metadata which can be used to manage updates.

otaUpdateId :: Text

The ID of the OTA update to be created.

targets :: NonEmpty Text

The devices targeted to receive OTA updates.

files :: NonEmpty OTAUpdateFile

The files to be streamed by the OTA update.

roleArn :: Text

The IAM role that grants Amazon Web Services IoT Core access to the Amazon S3, IoT jobs and Amazon Web Services Code Signing resources to create an OTA update job.

newCreateOTAUpdate

Arguments

:: Text
$sel:otaUpdateId:CreateOTAUpdate'

-> NonEmpty Text
$sel:targets:CreateOTAUpdate'

-> NonEmpty OTAUpdateFile
$sel:files:CreateOTAUpdate'

-> Text
$sel:roleArn:CreateOTAUpdate'

-> CreateOTAUpdate

Create a value of CreateOTAUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsJobAbortConfig:CreateOTAUpdate', createOTAUpdate_awsJobAbortConfig - The criteria that determine when and how a job abort takes place.

$sel:awsJobExecutionsRolloutConfig:CreateOTAUpdate', createOTAUpdate_awsJobExecutionsRolloutConfig - Configuration for the rollout of OTA updates.

$sel:protocols:CreateOTAUpdate', createOTAUpdate_protocols - The protocol used to transfer the OTA update image. Valid values are [HTTP], [MQTT], [HTTP, MQTT]. When both HTTP and MQTT are specified, the target device can choose the protocol.

$sel:awsJobPresignedUrlConfig:CreateOTAUpdate', createOTAUpdate_awsJobPresignedUrlConfig - Configuration information for pre-signed URLs.

$sel:additionalParameters:CreateOTAUpdate', createOTAUpdate_additionalParameters - A list of additional OTA update parameters which are name-value pairs.

$sel:awsJobTimeoutConfig:CreateOTAUpdate', createOTAUpdate_awsJobTimeoutConfig - Specifies the amount of time each device has to finish its execution of the job. A timer is started when the job execution status is set to IN_PROGRESS. If the job execution status is not set to another terminal state before the timer expires, it will be automatically set to TIMED_OUT.

$sel:description:CreateOTAUpdate', createOTAUpdate_description - The description of the OTA update.

$sel:targetSelection:CreateOTAUpdate', createOTAUpdate_targetSelection - Specifies whether the update will continue to run (CONTINUOUS), or will be complete after all the things specified as targets have completed the update (SNAPSHOT). If continuous, the update may also be run on a thing when a change is detected in a target. For example, an update will run on a thing when the thing is added to a target group, even after the update was completed by all things originally in the group. Valid values: CONTINUOUS | SNAPSHOT.

$sel:tags:CreateOTAUpdate', createOTAUpdate_tags - Metadata which can be used to manage updates.

$sel:otaUpdateId:CreateOTAUpdate', createOTAUpdate_otaUpdateId - The ID of the OTA update to be created.

$sel:targets:CreateOTAUpdate', createOTAUpdate_targets - The devices targeted to receive OTA updates.

$sel:files:CreateOTAUpdate', createOTAUpdate_files - The files to be streamed by the OTA update.

$sel:roleArn:CreateOTAUpdate', createOTAUpdate_roleArn - The IAM role that grants Amazon Web Services IoT Core access to the Amazon S3, IoT jobs and Amazon Web Services Code Signing resources to create an OTA update job.

Request Lenses

createOTAUpdate_awsJobAbortConfig :: Lens' CreateOTAUpdate (Maybe AwsJobAbortConfig)

The criteria that determine when and how a job abort takes place.

createOTAUpdate_awsJobExecutionsRolloutConfig :: Lens' CreateOTAUpdate (Maybe AwsJobExecutionsRolloutConfig)

Configuration for the rollout of OTA updates.

createOTAUpdate_protocols :: Lens' CreateOTAUpdate (Maybe (NonEmpty Protocol))

The protocol used to transfer the OTA update image. Valid values are [HTTP], [MQTT], [HTTP, MQTT]. When both HTTP and MQTT are specified, the target device can choose the protocol.

createOTAUpdate_awsJobPresignedUrlConfig :: Lens' CreateOTAUpdate (Maybe AwsJobPresignedUrlConfig)

Configuration information for pre-signed URLs.

createOTAUpdate_additionalParameters :: Lens' CreateOTAUpdate (Maybe (HashMap Text Text))

A list of additional OTA update parameters which are name-value pairs.

createOTAUpdate_awsJobTimeoutConfig :: Lens' CreateOTAUpdate (Maybe AwsJobTimeoutConfig)

Specifies the amount of time each device has to finish its execution of the job. A timer is started when the job execution status is set to IN_PROGRESS. If the job execution status is not set to another terminal state before the timer expires, it will be automatically set to TIMED_OUT.

createOTAUpdate_description :: Lens' CreateOTAUpdate (Maybe Text)

The description of the OTA update.

createOTAUpdate_targetSelection :: Lens' CreateOTAUpdate (Maybe TargetSelection)

Specifies whether the update will continue to run (CONTINUOUS), or will be complete after all the things specified as targets have completed the update (SNAPSHOT). If continuous, the update may also be run on a thing when a change is detected in a target. For example, an update will run on a thing when the thing is added to a target group, even after the update was completed by all things originally in the group. Valid values: CONTINUOUS | SNAPSHOT.

createOTAUpdate_tags :: Lens' CreateOTAUpdate (Maybe [Tag])

Metadata which can be used to manage updates.

createOTAUpdate_otaUpdateId :: Lens' CreateOTAUpdate Text

The ID of the OTA update to be created.

createOTAUpdate_targets :: Lens' CreateOTAUpdate (NonEmpty Text)

The devices targeted to receive OTA updates.

createOTAUpdate_files :: Lens' CreateOTAUpdate (NonEmpty OTAUpdateFile)

The files to be streamed by the OTA update.

createOTAUpdate_roleArn :: Lens' CreateOTAUpdate Text

The IAM role that grants Amazon Web Services IoT Core access to the Amazon S3, IoT jobs and Amazon Web Services Code Signing resources to create an OTA update job.

Destructuring the Response

data CreateOTAUpdateResponse

See: newCreateOTAUpdateResponse smart constructor.

Constructors

CreateOTAUpdateResponse'

Fields

awsIotJobId :: Maybe Text

The IoT job ID associated with the OTA update.

otaUpdateStatus :: Maybe OTAUpdateStatus

The OTA update status.

awsIotJobArn :: Maybe Text

The IoT job ARN associated with the OTA update.

otaUpdateId :: Maybe Text

The OTA update ID.

otaUpdateArn :: Maybe Text

The OTA update ARN.

httpStatus :: Int

The response's http status code.

newCreateOTAUpdateResponse

Arguments

:: Int
$sel:httpStatus:CreateOTAUpdateResponse'

-> CreateOTAUpdateResponse

Create a value of CreateOTAUpdateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsIotJobId:CreateOTAUpdateResponse', createOTAUpdateResponse_awsIotJobId - The IoT job ID associated with the OTA update.

$sel:otaUpdateStatus:CreateOTAUpdateResponse', createOTAUpdateResponse_otaUpdateStatus - The OTA update status.

$sel:awsIotJobArn:CreateOTAUpdateResponse', createOTAUpdateResponse_awsIotJobArn - The IoT job ARN associated with the OTA update.

$sel:otaUpdateId:CreateOTAUpdate', createOTAUpdateResponse_otaUpdateId - The OTA update ID.

$sel:otaUpdateArn:CreateOTAUpdateResponse', createOTAUpdateResponse_otaUpdateArn - The OTA update ARN.

$sel:httpStatus:CreateOTAUpdateResponse', createOTAUpdateResponse_httpStatus - The response's http status code.

Response Lenses

createOTAUpdateResponse_awsIotJobId :: Lens' CreateOTAUpdateResponse (Maybe Text)

The IoT job ID associated with the OTA update.

createOTAUpdateResponse_otaUpdateStatus :: Lens' CreateOTAUpdateResponse (Maybe OTAUpdateStatus)

The OTA update status.

createOTAUpdateResponse_awsIotJobArn :: Lens' CreateOTAUpdateResponse (Maybe Text)

The IoT job ARN associated with the OTA update.

createOTAUpdateResponse_otaUpdateId :: Lens' CreateOTAUpdateResponse (Maybe Text)

The OTA update ID.

createOTAUpdateResponse_otaUpdateArn :: Lens' CreateOTAUpdateResponse (Maybe Text)

The OTA update ARN.

createOTAUpdateResponse_httpStatus :: Lens' CreateOTAUpdateResponse Int

The response's http status code.