Network.AWS.IoT.CreateProvisioningClaim

Description

Creates a provisioning claim.

Requires permission to access the CreateProvisioningClaim action.

Creating a Request

data CreateProvisioningClaim

See: newCreateProvisioningClaim smart constructor.

Constructors

CreateProvisioningClaim'

Fields

templateName :: Text

The name of the provisioning template to use.

newCreateProvisioningClaim

Arguments

:: Text
$sel:templateName:CreateProvisioningClaim'

-> CreateProvisioningClaim

Create a value of CreateProvisioningClaim with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:CreateProvisioningClaim', createProvisioningClaim_templateName - The name of the provisioning template to use.

Request Lenses

createProvisioningClaim_templateName :: Lens' CreateProvisioningClaim Text

The name of the provisioning template to use.

Destructuring the Response

data CreateProvisioningClaimResponse

See: newCreateProvisioningClaimResponse smart constructor.

Constructors

CreateProvisioningClaimResponse'

Fields

expiration :: Maybe POSIX

The provisioning claim expiration time.

keyPair :: Maybe KeyPair

The provisioning claim key pair.

certificateId :: Maybe Text

The ID of the certificate.

certificatePem :: Maybe Text

The provisioning claim certificate.

httpStatus :: Int

The response's http status code.

newCreateProvisioningClaimResponse

Arguments

:: Int
$sel:httpStatus:CreateProvisioningClaimResponse'

-> CreateProvisioningClaimResponse

Create a value of CreateProvisioningClaimResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expiration:CreateProvisioningClaimResponse', createProvisioningClaimResponse_expiration - The provisioning claim expiration time.

$sel:keyPair:CreateProvisioningClaimResponse', createProvisioningClaimResponse_keyPair - The provisioning claim key pair.

$sel:certificateId:CreateProvisioningClaimResponse', createProvisioningClaimResponse_certificateId - The ID of the certificate.

$sel:certificatePem:CreateProvisioningClaimResponse', createProvisioningClaimResponse_certificatePem - The provisioning claim certificate.

$sel:httpStatus:CreateProvisioningClaimResponse', createProvisioningClaimResponse_httpStatus - The response's http status code.

Response Lenses

createProvisioningClaimResponse_expiration :: Lens' CreateProvisioningClaimResponse (Maybe UTCTime)

The provisioning claim expiration time.

createProvisioningClaimResponse_keyPair :: Lens' CreateProvisioningClaimResponse (Maybe KeyPair)

The provisioning claim key pair.

createProvisioningClaimResponse_certificateId :: Lens' CreateProvisioningClaimResponse (Maybe Text)

The ID of the certificate.

createProvisioningClaimResponse_certificatePem :: Lens' CreateProvisioningClaimResponse (Maybe Text)

The provisioning claim certificate.

createProvisioningClaimResponse_httpStatus :: Lens' CreateProvisioningClaimResponse Int

The response's http status code.