Network.AWS.IoT.CreateProvisioningTemplate

Description

Creates a fleet provisioning template.

Requires permission to access the CreateProvisioningTemplate action.

Creating a Request

data CreateProvisioningTemplate

See: newCreateProvisioningTemplate smart constructor.

Constructors

CreateProvisioningTemplate'

Fields

preProvisioningHook :: Maybe ProvisioningHook

Creates a pre-provisioning hook template.

enabled :: Maybe Bool

True to enable the fleet provisioning template, otherwise false.

tags :: Maybe [Tag]

Metadata which can be used to manage the fleet provisioning template.

For URI Request parameters use format: ...key1=value1&key2=value2...

For the CLI command-line parameter use format: &&tags "key1=value1&key2=value2..."

For the cli-input-json file use format: "tags": "key1=value1&key2=value2..."

description :: Maybe Text

The description of the fleet provisioning template.

templateName :: Text

The name of the fleet provisioning template.

templateBody :: Text

The JSON formatted contents of the fleet provisioning template.

provisioningRoleArn :: Text

The role ARN for the role associated with the fleet provisioning template. This IoT role grants permission to provision a device.

newCreateProvisioningTemplate

Arguments

:: Text
$sel:templateName:CreateProvisioningTemplate'

-> Text
$sel:templateBody:CreateProvisioningTemplate'

-> Text
$sel:provisioningRoleArn:CreateProvisioningTemplate'

-> CreateProvisioningTemplate

Create a value of CreateProvisioningTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:preProvisioningHook:CreateProvisioningTemplate', createProvisioningTemplate_preProvisioningHook - Creates a pre-provisioning hook template.

$sel:enabled:CreateProvisioningTemplate', createProvisioningTemplate_enabled - True to enable the fleet provisioning template, otherwise false.

$sel:tags:CreateProvisioningTemplate', createProvisioningTemplate_tags - Metadata which can be used to manage the fleet provisioning template.

For URI Request parameters use format: ...key1=value1&key2=value2...

For the CLI command-line parameter use format: &&tags "key1=value1&key2=value2..."

For the cli-input-json file use format: "tags": "key1=value1&key2=value2..."

$sel:description:CreateProvisioningTemplate', createProvisioningTemplate_description - The description of the fleet provisioning template.

$sel:templateName:CreateProvisioningTemplate', createProvisioningTemplate_templateName - The name of the fleet provisioning template.

$sel:templateBody:CreateProvisioningTemplate', createProvisioningTemplate_templateBody - The JSON formatted contents of the fleet provisioning template.

$sel:provisioningRoleArn:CreateProvisioningTemplate', createProvisioningTemplate_provisioningRoleArn - The role ARN for the role associated with the fleet provisioning template. This IoT role grants permission to provision a device.

Request Lenses

createProvisioningTemplate_preProvisioningHook :: Lens' CreateProvisioningTemplate (Maybe ProvisioningHook)

Creates a pre-provisioning hook template.

createProvisioningTemplate_enabled :: Lens' CreateProvisioningTemplate (Maybe Bool)

True to enable the fleet provisioning template, otherwise false.

createProvisioningTemplate_tags :: Lens' CreateProvisioningTemplate (Maybe [Tag])

Metadata which can be used to manage the fleet provisioning template.

For URI Request parameters use format: ...key1=value1&key2=value2...

For the CLI command-line parameter use format: &&tags "key1=value1&key2=value2..."

For the cli-input-json file use format: "tags": "key1=value1&key2=value2..."

createProvisioningTemplate_description :: Lens' CreateProvisioningTemplate (Maybe Text)

The description of the fleet provisioning template.

createProvisioningTemplate_templateName :: Lens' CreateProvisioningTemplate Text

The name of the fleet provisioning template.

createProvisioningTemplate_templateBody :: Lens' CreateProvisioningTemplate Text

The JSON formatted contents of the fleet provisioning template.

createProvisioningTemplate_provisioningRoleArn :: Lens' CreateProvisioningTemplate Text

The role ARN for the role associated with the fleet provisioning template. This IoT role grants permission to provision a device.

Destructuring the Response

data CreateProvisioningTemplateResponse

See: newCreateProvisioningTemplateResponse smart constructor.

Constructors

CreateProvisioningTemplateResponse'

Fields

templateName :: Maybe Text

The name of the fleet provisioning template.

defaultVersionId :: Maybe Int

The default version of the fleet provisioning template.

templateArn :: Maybe Text

The ARN that identifies the provisioning template.

httpStatus :: Int

The response's http status code.

newCreateProvisioningTemplateResponse

Arguments

:: Int
$sel:httpStatus:CreateProvisioningTemplateResponse'

-> CreateProvisioningTemplateResponse

Create a value of CreateProvisioningTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:CreateProvisioningTemplate', createProvisioningTemplateResponse_templateName - The name of the fleet provisioning template.

$sel:defaultVersionId:CreateProvisioningTemplateResponse', createProvisioningTemplateResponse_defaultVersionId - The default version of the fleet provisioning template.

$sel:templateArn:CreateProvisioningTemplateResponse', createProvisioningTemplateResponse_templateArn - The ARN that identifies the provisioning template.

$sel:httpStatus:CreateProvisioningTemplateResponse', createProvisioningTemplateResponse_httpStatus - The response's http status code.

Response Lenses

createProvisioningTemplateResponse_templateName :: Lens' CreateProvisioningTemplateResponse (Maybe Text)

The name of the fleet provisioning template.

createProvisioningTemplateResponse_defaultVersionId :: Lens' CreateProvisioningTemplateResponse (Maybe Int)

The default version of the fleet provisioning template.

createProvisioningTemplateResponse_templateArn :: Lens' CreateProvisioningTemplateResponse (Maybe Text)

The ARN that identifies the provisioning template.

createProvisioningTemplateResponse_httpStatus :: Lens' CreateProvisioningTemplateResponse Int

The response's http status code.