Network.AWS.IoT.CreateProvisioningTemplateVersion

Description

Creates a new version of a fleet provisioning template.

Requires permission to access the CreateProvisioningTemplateVersion action.

Creating a Request

data CreateProvisioningTemplateVersion

See: newCreateProvisioningTemplateVersion smart constructor.

Constructors

CreateProvisioningTemplateVersion'

Fields

setAsDefault :: Maybe Bool

Sets a fleet provision template version as the default version.

templateName :: Text

The name of the fleet provisioning template.

templateBody :: Text

The JSON formatted contents of the fleet provisioning template.

newCreateProvisioningTemplateVersion

Arguments

:: Text
$sel:templateName:CreateProvisioningTemplateVersion'

-> Text
$sel:templateBody:CreateProvisioningTemplateVersion'

-> CreateProvisioningTemplateVersion

Create a value of CreateProvisioningTemplateVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:setAsDefault:CreateProvisioningTemplateVersion', createProvisioningTemplateVersion_setAsDefault - Sets a fleet provision template version as the default version.

$sel:templateName:CreateProvisioningTemplateVersion', createProvisioningTemplateVersion_templateName - The name of the fleet provisioning template.

$sel:templateBody:CreateProvisioningTemplateVersion', createProvisioningTemplateVersion_templateBody - The JSON formatted contents of the fleet provisioning template.

Request Lenses

createProvisioningTemplateVersion_setAsDefault :: Lens' CreateProvisioningTemplateVersion (Maybe Bool)

Sets a fleet provision template version as the default version.

createProvisioningTemplateVersion_templateName :: Lens' CreateProvisioningTemplateVersion Text

The name of the fleet provisioning template.

createProvisioningTemplateVersion_templateBody :: Lens' CreateProvisioningTemplateVersion Text

The JSON formatted contents of the fleet provisioning template.

Destructuring the Response

data CreateProvisioningTemplateVersionResponse

See: newCreateProvisioningTemplateVersionResponse smart constructor.

Constructors

CreateProvisioningTemplateVersionResponse'

Fields

versionId :: Maybe Int

The version of the fleet provisioning template.

templateName :: Maybe Text

The name of the fleet provisioning template.

templateArn :: Maybe Text

The ARN that identifies the provisioning template.

isDefaultVersion :: Maybe Bool

True if the fleet provisioning template version is the default version, otherwise false.

httpStatus :: Int

The response's http status code.

newCreateProvisioningTemplateVersionResponse

Arguments

:: Int
$sel:httpStatus:CreateProvisioningTemplateVersionResponse'

-> CreateProvisioningTemplateVersionResponse

Create a value of CreateProvisioningTemplateVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionId:CreateProvisioningTemplateVersionResponse', createProvisioningTemplateVersionResponse_versionId - The version of the fleet provisioning template.

$sel:templateName:CreateProvisioningTemplateVersion', createProvisioningTemplateVersionResponse_templateName - The name of the fleet provisioning template.

$sel:templateArn:CreateProvisioningTemplateVersionResponse', createProvisioningTemplateVersionResponse_templateArn - The ARN that identifies the provisioning template.

$sel:isDefaultVersion:CreateProvisioningTemplateVersionResponse', createProvisioningTemplateVersionResponse_isDefaultVersion - True if the fleet provisioning template version is the default version, otherwise false.

$sel:httpStatus:CreateProvisioningTemplateVersionResponse', createProvisioningTemplateVersionResponse_httpStatus - The response's http status code.

Response Lenses

createProvisioningTemplateVersionResponse_versionId :: Lens' CreateProvisioningTemplateVersionResponse (Maybe Int)

The version of the fleet provisioning template.

createProvisioningTemplateVersionResponse_templateName :: Lens' CreateProvisioningTemplateVersionResponse (Maybe Text)

The name of the fleet provisioning template.

createProvisioningTemplateVersionResponse_templateArn :: Lens' CreateProvisioningTemplateVersionResponse (Maybe Text)

The ARN that identifies the provisioning template.

createProvisioningTemplateVersionResponse_isDefaultVersion :: Lens' CreateProvisioningTemplateVersionResponse (Maybe Bool)

True if the fleet provisioning template version is the default version, otherwise false.

createProvisioningTemplateVersionResponse_httpStatus :: Lens' CreateProvisioningTemplateVersionResponse Int

The response's http status code.