Network.AWS.IoT.CreateSecurityProfile

Description

Creates a Device Defender security profile.

Requires permission to access the CreateSecurityProfile action.

Creating a Request

data CreateSecurityProfile

See: newCreateSecurityProfile smart constructor.

Constructors

CreateSecurityProfile'

Fields

alertTargets :: Maybe (HashMap AlertTargetType AlertTarget)

Specifies the destinations to which alerts are sent. (Alerts are always sent to the console.) Alerts are generated when a device (thing) violates a behavior.

additionalMetricsToRetainV2 :: Maybe [MetricToRetain]

A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the profile's behaviors, but it is also retained for any metric specified here. Can be used with custom metrics; cannot be used with dimensions.

behaviors :: Maybe [Behavior]

Specifies the behaviors that, when violated by a device (thing), cause an alert.

additionalMetricsToRetain :: Maybe [Text]

/Please use CreateSecurityProfileRequest$additionalMetricsToRetainV2 instead./

A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the profile's behaviors, but it is also retained for any metric specified here. Can be used with custom metrics; cannot be used with dimensions.

securityProfileDescription :: Maybe Text

A description of the security profile.

tags :: Maybe [Tag]

Metadata that can be used to manage the security profile.

securityProfileName :: Text

The name you are giving to the security profile.

newCreateSecurityProfile

Arguments

:: Text
$sel:securityProfileName:CreateSecurityProfile'

-> CreateSecurityProfile

Create a value of CreateSecurityProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alertTargets:CreateSecurityProfile', createSecurityProfile_alertTargets - Specifies the destinations to which alerts are sent. (Alerts are always sent to the console.) Alerts are generated when a device (thing) violates a behavior.

$sel:additionalMetricsToRetainV2:CreateSecurityProfile', createSecurityProfile_additionalMetricsToRetainV2 - A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the profile's behaviors, but it is also retained for any metric specified here. Can be used with custom metrics; cannot be used with dimensions.

$sel:behaviors:CreateSecurityProfile', createSecurityProfile_behaviors - Specifies the behaviors that, when violated by a device (thing), cause an alert.

$sel:additionalMetricsToRetain:CreateSecurityProfile', createSecurityProfile_additionalMetricsToRetain - /Please use CreateSecurityProfileRequest$additionalMetricsToRetainV2 instead./

A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the profile's behaviors, but it is also retained for any metric specified here. Can be used with custom metrics; cannot be used with dimensions.

$sel:securityProfileDescription:CreateSecurityProfile', createSecurityProfile_securityProfileDescription - A description of the security profile.

$sel:tags:CreateSecurityProfile', createSecurityProfile_tags - Metadata that can be used to manage the security profile.

$sel:securityProfileName:CreateSecurityProfile', createSecurityProfile_securityProfileName - The name you are giving to the security profile.

Request Lenses

createSecurityProfile_alertTargets :: Lens' CreateSecurityProfile (Maybe (HashMap AlertTargetType AlertTarget))

Specifies the destinations to which alerts are sent. (Alerts are always sent to the console.) Alerts are generated when a device (thing) violates a behavior.

createSecurityProfile_additionalMetricsToRetainV2 :: Lens' CreateSecurityProfile (Maybe [MetricToRetain])

A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the profile's behaviors, but it is also retained for any metric specified here. Can be used with custom metrics; cannot be used with dimensions.

createSecurityProfile_behaviors :: Lens' CreateSecurityProfile (Maybe [Behavior])

Specifies the behaviors that, when violated by a device (thing), cause an alert.

createSecurityProfile_additionalMetricsToRetain :: Lens' CreateSecurityProfile (Maybe [Text])

/Please use CreateSecurityProfileRequest$additionalMetricsToRetainV2 instead./

A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the profile's behaviors, but it is also retained for any metric specified here. Can be used with custom metrics; cannot be used with dimensions.

createSecurityProfile_securityProfileDescription :: Lens' CreateSecurityProfile (Maybe Text)

A description of the security profile.

createSecurityProfile_tags :: Lens' CreateSecurityProfile (Maybe [Tag])

Metadata that can be used to manage the security profile.

createSecurityProfile_securityProfileName :: Lens' CreateSecurityProfile Text

The name you are giving to the security profile.

Destructuring the Response

data CreateSecurityProfileResponse

See: newCreateSecurityProfileResponse smart constructor.

Constructors

CreateSecurityProfileResponse'

Fields

securityProfileName :: Maybe Text

The name you gave to the security profile.

securityProfileArn :: Maybe Text

The ARN of the security profile.

httpStatus :: Int

The response's http status code.

newCreateSecurityProfileResponse

Arguments

:: Int
$sel:httpStatus:CreateSecurityProfileResponse'

-> CreateSecurityProfileResponse

Create a value of CreateSecurityProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityProfileName:CreateSecurityProfile', createSecurityProfileResponse_securityProfileName - The name you gave to the security profile.

$sel:securityProfileArn:CreateSecurityProfileResponse', createSecurityProfileResponse_securityProfileArn - The ARN of the security profile.

$sel:httpStatus:CreateSecurityProfileResponse', createSecurityProfileResponse_httpStatus - The response's http status code.

Response Lenses

createSecurityProfileResponse_securityProfileName :: Lens' CreateSecurityProfileResponse (Maybe Text)

The name you gave to the security profile.

createSecurityProfileResponse_securityProfileArn :: Lens' CreateSecurityProfileResponse (Maybe Text)

The ARN of the security profile.

createSecurityProfileResponse_httpStatus :: Lens' CreateSecurityProfileResponse Int

The response's http status code.