Network.AWS.IoT.CreateThingGroup

Description

Create a thing group.

This is a control plane operation. See Authorization for information about authorizing control plane actions.

Requires permission to access the CreateThingGroup action.

Creating a Request

data CreateThingGroup

See: newCreateThingGroup smart constructor.

Constructors

CreateThingGroup'

Fields

parentGroupName :: Maybe Text

The name of the parent thing group.

tags :: Maybe [Tag]

Metadata which can be used to manage the thing group.

thingGroupProperties :: Maybe ThingGroupProperties

The thing group properties.

thingGroupName :: Text

The thing group name to create.

newCreateThingGroup

Arguments

:: Text
$sel:thingGroupName:CreateThingGroup'

-> CreateThingGroup

Create a value of CreateThingGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parentGroupName:CreateThingGroup', createThingGroup_parentGroupName - The name of the parent thing group.

$sel:tags:CreateThingGroup', createThingGroup_tags - Metadata which can be used to manage the thing group.

$sel:thingGroupProperties:CreateThingGroup', createThingGroup_thingGroupProperties - The thing group properties.

$sel:thingGroupName:CreateThingGroup', createThingGroup_thingGroupName - The thing group name to create.

Request Lenses

createThingGroup_parentGroupName :: Lens' CreateThingGroup (Maybe Text)

The name of the parent thing group.

createThingGroup_tags :: Lens' CreateThingGroup (Maybe [Tag])

Metadata which can be used to manage the thing group.

createThingGroup_thingGroupProperties :: Lens' CreateThingGroup (Maybe ThingGroupProperties)

The thing group properties.

createThingGroup_thingGroupName :: Lens' CreateThingGroup Text

The thing group name to create.

Destructuring the Response

data CreateThingGroupResponse

See: newCreateThingGroupResponse smart constructor.

Constructors

CreateThingGroupResponse'

Fields

thingGroupArn :: Maybe Text

The thing group ARN.

thingGroupName :: Maybe Text

The thing group name.

thingGroupId :: Maybe Text

The thing group ID.

httpStatus :: Int

The response's http status code.

newCreateThingGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateThingGroupResponse'

-> CreateThingGroupResponse

Create a value of CreateThingGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thingGroupArn:CreateThingGroupResponse', createThingGroupResponse_thingGroupArn - The thing group ARN.

$sel:thingGroupName:CreateThingGroup', createThingGroupResponse_thingGroupName - The thing group name.

$sel:thingGroupId:CreateThingGroupResponse', createThingGroupResponse_thingGroupId - The thing group ID.

$sel:httpStatus:CreateThingGroupResponse', createThingGroupResponse_httpStatus - The response's http status code.

Response Lenses

createThingGroupResponse_thingGroupArn :: Lens' CreateThingGroupResponse (Maybe Text)

The thing group ARN.

createThingGroupResponse_thingGroupName :: Lens' CreateThingGroupResponse (Maybe Text)

The thing group name.

createThingGroupResponse_thingGroupId :: Lens' CreateThingGroupResponse (Maybe Text)

The thing group ID.

createThingGroupResponse_httpStatus :: Lens' CreateThingGroupResponse Int

The response's http status code.