Network.AWS.IoT.CreateTopicRuleDestination

Description

Creates a topic rule destination. The destination must be confirmed prior to use.

Requires permission to access the CreateTopicRuleDestination action.

Creating a Request

data CreateTopicRuleDestination

See: newCreateTopicRuleDestination smart constructor.

Constructors

CreateTopicRuleDestination'

Fields

destinationConfiguration :: TopicRuleDestinationConfiguration

The topic rule destination configuration.

newCreateTopicRuleDestination

Arguments

:: TopicRuleDestinationConfiguration
$sel:destinationConfiguration:CreateTopicRuleDestination'

-> CreateTopicRuleDestination

Create a value of CreateTopicRuleDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationConfiguration:CreateTopicRuleDestination', createTopicRuleDestination_destinationConfiguration - The topic rule destination configuration.

Request Lenses

createTopicRuleDestination_destinationConfiguration :: Lens' CreateTopicRuleDestination TopicRuleDestinationConfiguration

The topic rule destination configuration.

Destructuring the Response

data CreateTopicRuleDestinationResponse

See: newCreateTopicRuleDestinationResponse smart constructor.

Constructors

CreateTopicRuleDestinationResponse'

Fields

topicRuleDestination :: Maybe TopicRuleDestination

The topic rule destination.

httpStatus :: Int

The response's http status code.

newCreateTopicRuleDestinationResponse

Arguments

:: Int
$sel:httpStatus:CreateTopicRuleDestinationResponse'

-> CreateTopicRuleDestinationResponse

Create a value of CreateTopicRuleDestinationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:topicRuleDestination:CreateTopicRuleDestinationResponse', createTopicRuleDestinationResponse_topicRuleDestination - The topic rule destination.

$sel:httpStatus:CreateTopicRuleDestinationResponse', createTopicRuleDestinationResponse_httpStatus - The response's http status code.

Response Lenses

createTopicRuleDestinationResponse_topicRuleDestination :: Lens' CreateTopicRuleDestinationResponse (Maybe TopicRuleDestination)

The topic rule destination.

createTopicRuleDestinationResponse_httpStatus :: Lens' CreateTopicRuleDestinationResponse Int

The response's http status code.