Network.AWS.IoT.DeleteDimension

Description

Removes the specified dimension from your Amazon Web Services accounts.

Requires permission to access the DeleteDimension action.

Creating a Request

data DeleteDimension

See: newDeleteDimension smart constructor.

Constructors

DeleteDimension'

Fields

name :: Text

The unique identifier for the dimension that you want to delete.

newDeleteDimension

Arguments

:: Text
$sel:name:DeleteDimension'

-> DeleteDimension

Create a value of DeleteDimension with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DeleteDimension', deleteDimension_name - The unique identifier for the dimension that you want to delete.

Request Lenses

deleteDimension_name :: Lens' DeleteDimension Text

The unique identifier for the dimension that you want to delete.

Destructuring the Response

data DeleteDimensionResponse

See: newDeleteDimensionResponse smart constructor.

Constructors

DeleteDimensionResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteDimensionResponse

Arguments

:: Int
$sel:httpStatus:DeleteDimensionResponse'

-> DeleteDimensionResponse

Create a value of DeleteDimensionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteDimensionResponse', deleteDimensionResponse_httpStatus - The response's http status code.

Response Lenses

deleteDimensionResponse_httpStatus :: Lens' DeleteDimensionResponse Int

The response's http status code.