Network.AWS.IoT.DeleteProvisioningTemplate

Description

Deletes a fleet provisioning template.

Requires permission to access the DeleteProvisioningTemplate action.

Creating a Request

data DeleteProvisioningTemplate

See: newDeleteProvisioningTemplate smart constructor.

Constructors

DeleteProvisioningTemplate'

Fields

templateName :: Text

The name of the fleet provision template to delete.

newDeleteProvisioningTemplate

Arguments

:: Text
$sel:templateName:DeleteProvisioningTemplate'

-> DeleteProvisioningTemplate

Create a value of DeleteProvisioningTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:DeleteProvisioningTemplate', deleteProvisioningTemplate_templateName - The name of the fleet provision template to delete.

Request Lenses

deleteProvisioningTemplate_templateName :: Lens' DeleteProvisioningTemplate Text

The name of the fleet provision template to delete.

Destructuring the Response

data DeleteProvisioningTemplateResponse

See: newDeleteProvisioningTemplateResponse smart constructor.

Constructors

DeleteProvisioningTemplateResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteProvisioningTemplateResponse

Arguments

:: Int
$sel:httpStatus:DeleteProvisioningTemplateResponse'

-> DeleteProvisioningTemplateResponse

Create a value of DeleteProvisioningTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteProvisioningTemplateResponse', deleteProvisioningTemplateResponse_httpStatus - The response's http status code.

Response Lenses

deleteProvisioningTemplateResponse_httpStatus :: Lens' DeleteProvisioningTemplateResponse Int

The response's http status code.