Network.AWS.IoT.DescribeAccountAuditConfiguration

Description

Gets information about the Device Defender audit settings for this account. Settings include how audit notifications are sent and which audit checks are enabled or disabled.

Requires permission to access the DescribeAccountAuditConfiguration action.

Creating a Request

data DescribeAccountAuditConfiguration

See: newDescribeAccountAuditConfiguration smart constructor.

Constructors

DescribeAccountAuditConfiguration'

newDescribeAccountAuditConfiguration :: DescribeAccountAuditConfiguration

Create a value of DescribeAccountAuditConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data DescribeAccountAuditConfigurationResponse

See: newDescribeAccountAuditConfigurationResponse smart constructor.

Constructors

DescribeAccountAuditConfigurationResponse'

Fields

roleArn :: Maybe Text

The ARN of the role that grants permission to IoT to access information about your devices, policies, certificates, and other items as required when performing an audit.

On the first call to UpdateAccountAuditConfiguration, this parameter is required.

auditCheckConfigurations :: Maybe (HashMap Text AuditCheckConfiguration)

Which audit checks are enabled and disabled for this account.

auditNotificationTargetConfigurations :: Maybe (HashMap AuditNotificationType AuditNotificationTarget)

Information about the targets to which audit notifications are sent for this account.

httpStatus :: Int

The response's http status code.

newDescribeAccountAuditConfigurationResponse

Arguments

:: Int
$sel:httpStatus:DescribeAccountAuditConfigurationResponse'

-> DescribeAccountAuditConfigurationResponse

Create a value of DescribeAccountAuditConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleArn:DescribeAccountAuditConfigurationResponse', describeAccountAuditConfigurationResponse_roleArn - The ARN of the role that grants permission to IoT to access information about your devices, policies, certificates, and other items as required when performing an audit.

On the first call to UpdateAccountAuditConfiguration, this parameter is required.

$sel:auditCheckConfigurations:DescribeAccountAuditConfigurationResponse', describeAccountAuditConfigurationResponse_auditCheckConfigurations - Which audit checks are enabled and disabled for this account.

$sel:auditNotificationTargetConfigurations:DescribeAccountAuditConfigurationResponse', describeAccountAuditConfigurationResponse_auditNotificationTargetConfigurations - Information about the targets to which audit notifications are sent for this account.

$sel:httpStatus:DescribeAccountAuditConfigurationResponse', describeAccountAuditConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

describeAccountAuditConfigurationResponse_roleArn :: Lens' DescribeAccountAuditConfigurationResponse (Maybe Text)

The ARN of the role that grants permission to IoT to access information about your devices, policies, certificates, and other items as required when performing an audit.

On the first call to UpdateAccountAuditConfiguration, this parameter is required.

describeAccountAuditConfigurationResponse_auditCheckConfigurations :: Lens' DescribeAccountAuditConfigurationResponse (Maybe (HashMap Text AuditCheckConfiguration))

Which audit checks are enabled and disabled for this account.

describeAccountAuditConfigurationResponse_auditNotificationTargetConfigurations :: Lens' DescribeAccountAuditConfigurationResponse (Maybe (HashMap AuditNotificationType AuditNotificationTarget))

Information about the targets to which audit notifications are sent for this account.

describeAccountAuditConfigurationResponse_httpStatus :: Lens' DescribeAccountAuditConfigurationResponse Int

The response's http status code.