Network.AWS.IoT.DescribeAuditMitigationActionsTask

Description

Gets information about an audit mitigation task that is used to apply mitigation actions to a set of audit findings. Properties include the actions being applied, the audit checks to which they're being applied, the task status, and aggregated task statistics.

Creating a Request

data DescribeAuditMitigationActionsTask

See: newDescribeAuditMitigationActionsTask smart constructor.

Constructors

DescribeAuditMitigationActionsTask'

Fields

taskId :: Text

The unique identifier for the audit mitigation task.

newDescribeAuditMitigationActionsTask

Arguments

:: Text
$sel:taskId:DescribeAuditMitigationActionsTask'

-> DescribeAuditMitigationActionsTask

Create a value of DescribeAuditMitigationActionsTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskId:DescribeAuditMitigationActionsTask', describeAuditMitigationActionsTask_taskId - The unique identifier for the audit mitigation task.

Request Lenses

describeAuditMitigationActionsTask_taskId :: Lens' DescribeAuditMitigationActionsTask Text

The unique identifier for the audit mitigation task.

Destructuring the Response

data DescribeAuditMitigationActionsTaskResponse

See: newDescribeAuditMitigationActionsTaskResponse smart constructor.

Constructors

DescribeAuditMitigationActionsTaskResponse'

Fields

auditCheckToActionsMapping :: Maybe (HashMap Text (NonEmpty Text))

Specifies the mitigation actions that should be applied to specific audit checks.

taskStatistics :: Maybe (HashMap Text TaskStatisticsForAuditCheck)

Aggregate counts of the results when the mitigation tasks were applied to the findings for this audit mitigation actions task.

startTime :: Maybe POSIX

The date and time when the task was started.

endTime :: Maybe POSIX

The date and time when the task was completed or canceled.

target :: Maybe AuditMitigationActionsTaskTarget

Identifies the findings to which the mitigation actions are applied. This can be by audit checks, by audit task, or a set of findings.

taskStatus :: Maybe AuditMitigationActionsTaskStatus

The current status of the task.

actionsDefinition :: Maybe [MitigationAction]

Specifies the mitigation actions and their parameters that are applied as part of this task.

httpStatus :: Int

The response's http status code.

newDescribeAuditMitigationActionsTaskResponse

Arguments

:: Int
$sel:httpStatus:DescribeAuditMitigationActionsTaskResponse'

-> DescribeAuditMitigationActionsTaskResponse

Create a value of DescribeAuditMitigationActionsTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:auditCheckToActionsMapping:DescribeAuditMitigationActionsTaskResponse', describeAuditMitigationActionsTaskResponse_auditCheckToActionsMapping - Specifies the mitigation actions that should be applied to specific audit checks.

$sel:taskStatistics:DescribeAuditMitigationActionsTaskResponse', describeAuditMitigationActionsTaskResponse_taskStatistics - Aggregate counts of the results when the mitigation tasks were applied to the findings for this audit mitigation actions task.

$sel:startTime:DescribeAuditMitigationActionsTaskResponse', describeAuditMitigationActionsTaskResponse_startTime - The date and time when the task was started.

$sel:endTime:DescribeAuditMitigationActionsTaskResponse', describeAuditMitigationActionsTaskResponse_endTime - The date and time when the task was completed or canceled.

$sel:target:DescribeAuditMitigationActionsTaskResponse', describeAuditMitigationActionsTaskResponse_target - Identifies the findings to which the mitigation actions are applied. This can be by audit checks, by audit task, or a set of findings.

$sel:taskStatus:DescribeAuditMitigationActionsTaskResponse', describeAuditMitigationActionsTaskResponse_taskStatus - The current status of the task.

$sel:actionsDefinition:DescribeAuditMitigationActionsTaskResponse', describeAuditMitigationActionsTaskResponse_actionsDefinition - Specifies the mitigation actions and their parameters that are applied as part of this task.

$sel:httpStatus:DescribeAuditMitigationActionsTaskResponse', describeAuditMitigationActionsTaskResponse_httpStatus - The response's http status code.

Response Lenses

describeAuditMitigationActionsTaskResponse_auditCheckToActionsMapping :: Lens' DescribeAuditMitigationActionsTaskResponse (Maybe (HashMap Text (NonEmpty Text)))

Specifies the mitigation actions that should be applied to specific audit checks.

describeAuditMitigationActionsTaskResponse_taskStatistics :: Lens' DescribeAuditMitigationActionsTaskResponse (Maybe (HashMap Text TaskStatisticsForAuditCheck))

Aggregate counts of the results when the mitigation tasks were applied to the findings for this audit mitigation actions task.

describeAuditMitigationActionsTaskResponse_startTime :: Lens' DescribeAuditMitigationActionsTaskResponse (Maybe UTCTime)

The date and time when the task was started.

describeAuditMitigationActionsTaskResponse_endTime :: Lens' DescribeAuditMitigationActionsTaskResponse (Maybe UTCTime)

The date and time when the task was completed or canceled.

describeAuditMitigationActionsTaskResponse_target :: Lens' DescribeAuditMitigationActionsTaskResponse (Maybe AuditMitigationActionsTaskTarget)

Identifies the findings to which the mitigation actions are applied. This can be by audit checks, by audit task, or a set of findings.

describeAuditMitigationActionsTaskResponse_taskStatus :: Lens' DescribeAuditMitigationActionsTaskResponse (Maybe AuditMitigationActionsTaskStatus)

The current status of the task.

describeAuditMitigationActionsTaskResponse_actionsDefinition :: Lens' DescribeAuditMitigationActionsTaskResponse (Maybe [MitigationAction])

Specifies the mitigation actions and their parameters that are applied as part of this task.

describeAuditMitigationActionsTaskResponse_httpStatus :: Lens' DescribeAuditMitigationActionsTaskResponse Int

The response's http status code.