Network.AWS.IoT.DescribeCustomMetric

Description

Gets information about a Device Defender detect custom metric.

Requires permission to access the DescribeCustomMetric action.

Creating a Request

data DescribeCustomMetric

See: newDescribeCustomMetric smart constructor.

Constructors

DescribeCustomMetric'

Fields

metricName :: Text

The name of the custom metric.

newDescribeCustomMetric

Arguments

:: Text
$sel:metricName:DescribeCustomMetric'

-> DescribeCustomMetric

Create a value of DescribeCustomMetric with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricName:DescribeCustomMetric', describeCustomMetric_metricName - The name of the custom metric.

Request Lenses

describeCustomMetric_metricName :: Lens' DescribeCustomMetric Text

The name of the custom metric.

Destructuring the Response

data DescribeCustomMetricResponse

See: newDescribeCustomMetricResponse smart constructor.

Constructors

DescribeCustomMetricResponse'

Fields

metricType :: Maybe CustomMetricType

The type of the custom metric. Types include string-list, ip-address-list, number-list, and number.

lastModifiedDate :: Maybe POSIX

The time the custom metric was last modified in milliseconds since epoch.

metricName :: Maybe Text

The name of the custom metric.

displayName :: Maybe Text

Field represents a friendly name in the console for the custom metric; doesn't have to be unique. Don't use this name as the metric identifier in the device metric report. Can be updated.

creationDate :: Maybe POSIX

The creation date of the custom metric in milliseconds since epoch.

metricArn :: Maybe Text

The Amazon Resource Number (ARN) of the custom metric.

httpStatus :: Int

The response's http status code.

newDescribeCustomMetricResponse

Arguments

:: Int
$sel:httpStatus:DescribeCustomMetricResponse'

-> DescribeCustomMetricResponse

Create a value of DescribeCustomMetricResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricType:DescribeCustomMetricResponse', describeCustomMetricResponse_metricType - The type of the custom metric. Types include string-list, ip-address-list, number-list, and number.

$sel:lastModifiedDate:DescribeCustomMetricResponse', describeCustomMetricResponse_lastModifiedDate - The time the custom metric was last modified in milliseconds since epoch.

$sel:metricName:DescribeCustomMetric', describeCustomMetricResponse_metricName - The name of the custom metric.

$sel:displayName:DescribeCustomMetricResponse', describeCustomMetricResponse_displayName - Field represents a friendly name in the console for the custom metric; doesn't have to be unique. Don't use this name as the metric identifier in the device metric report. Can be updated.

$sel:creationDate:DescribeCustomMetricResponse', describeCustomMetricResponse_creationDate - The creation date of the custom metric in milliseconds since epoch.

$sel:metricArn:DescribeCustomMetricResponse', describeCustomMetricResponse_metricArn - The Amazon Resource Number (ARN) of the custom metric.

$sel:httpStatus:DescribeCustomMetricResponse', describeCustomMetricResponse_httpStatus - The response's http status code.

Response Lenses

describeCustomMetricResponse_metricType :: Lens' DescribeCustomMetricResponse (Maybe CustomMetricType)

The type of the custom metric. Types include string-list, ip-address-list, number-list, and number.

describeCustomMetricResponse_lastModifiedDate :: Lens' DescribeCustomMetricResponse (Maybe UTCTime)

The time the custom metric was last modified in milliseconds since epoch.

describeCustomMetricResponse_metricName :: Lens' DescribeCustomMetricResponse (Maybe Text)

The name of the custom metric.

describeCustomMetricResponse_displayName :: Lens' DescribeCustomMetricResponse (Maybe Text)

Field represents a friendly name in the console for the custom metric; doesn't have to be unique. Don't use this name as the metric identifier in the device metric report. Can be updated.

describeCustomMetricResponse_creationDate :: Lens' DescribeCustomMetricResponse (Maybe UTCTime)

The creation date of the custom metric in milliseconds since epoch.

describeCustomMetricResponse_metricArn :: Lens' DescribeCustomMetricResponse (Maybe Text)

The Amazon Resource Number (ARN) of the custom metric.

describeCustomMetricResponse_httpStatus :: Lens' DescribeCustomMetricResponse Int

The response's http status code.