Network.AWS.IoT.DescribeDimension

Description

Provides details about a dimension that is defined in your Amazon Web Services accounts.

Requires permission to access the DescribeDimension action.

Creating a Request

data DescribeDimension

See: newDescribeDimension smart constructor.

Constructors

DescribeDimension'

Fields

name :: Text

The unique identifier for the dimension.

newDescribeDimension

Arguments

:: Text
$sel:name:DescribeDimension'

-> DescribeDimension

Create a value of DescribeDimension with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DescribeDimension', describeDimension_name - The unique identifier for the dimension.

Request Lenses

describeDimension_name :: Lens' DescribeDimension Text

The unique identifier for the dimension.

Destructuring the Response

data DescribeDimensionResponse

See: newDescribeDimensionResponse smart constructor.

Constructors

DescribeDimensionResponse'

Fields

lastModifiedDate :: Maybe POSIX

The date the dimension was last modified.

arn :: Maybe Text

The Amazon Resource Name (ARN) for the dimension.

stringValues :: Maybe (NonEmpty Text)

The value or list of values used to scope the dimension. For example, for topic filters, this is the pattern used to match the MQTT topic name.

name :: Maybe Text

The unique identifier for the dimension.

creationDate :: Maybe POSIX

The date the dimension was created.

type' :: Maybe DimensionType

The type of the dimension.

httpStatus :: Int

The response's http status code.

newDescribeDimensionResponse

Arguments

:: Int
$sel:httpStatus:DescribeDimensionResponse'

-> DescribeDimensionResponse

Create a value of DescribeDimensionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:DescribeDimensionResponse', describeDimensionResponse_lastModifiedDate - The date the dimension was last modified.

$sel:arn:DescribeDimensionResponse', describeDimensionResponse_arn - The Amazon Resource Name (ARN) for the dimension.

$sel:stringValues:DescribeDimensionResponse', describeDimensionResponse_stringValues - The value or list of values used to scope the dimension. For example, for topic filters, this is the pattern used to match the MQTT topic name.

$sel:name:DescribeDimension', describeDimensionResponse_name - The unique identifier for the dimension.

$sel:creationDate:DescribeDimensionResponse', describeDimensionResponse_creationDate - The date the dimension was created.

$sel:type':DescribeDimensionResponse', describeDimensionResponse_type - The type of the dimension.

$sel:httpStatus:DescribeDimensionResponse', describeDimensionResponse_httpStatus - The response's http status code.

Response Lenses

describeDimensionResponse_lastModifiedDate :: Lens' DescribeDimensionResponse (Maybe UTCTime)

The date the dimension was last modified.

describeDimensionResponse_arn :: Lens' DescribeDimensionResponse (Maybe Text)

The Amazon Resource Name (ARN) for the dimension.

describeDimensionResponse_stringValues :: Lens' DescribeDimensionResponse (Maybe (NonEmpty Text))

The value or list of values used to scope the dimension. For example, for topic filters, this is the pattern used to match the MQTT topic name.

describeDimensionResponse_name :: Lens' DescribeDimensionResponse (Maybe Text)

The unique identifier for the dimension.

describeDimensionResponse_creationDate :: Lens' DescribeDimensionResponse (Maybe UTCTime)

The date the dimension was created.

describeDimensionResponse_type :: Lens' DescribeDimensionResponse (Maybe DimensionType)

The type of the dimension.

describeDimensionResponse_httpStatus :: Lens' DescribeDimensionResponse Int

The response's http status code.