Network.AWS.IoT.DescribeDomainConfiguration

Description

Gets summary information about a domain configuration.

Requires permission to access the DescribeDomainConfiguration action.

Creating a Request

data DescribeDomainConfiguration

See: newDescribeDomainConfiguration smart constructor.

Constructors

DescribeDomainConfiguration'

Fields

domainConfigurationName :: Text

The name of the domain configuration.

newDescribeDomainConfiguration

Arguments

:: Text
$sel:domainConfigurationName:DescribeDomainConfiguration'

-> DescribeDomainConfiguration

Create a value of DescribeDomainConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainConfigurationName:DescribeDomainConfiguration', describeDomainConfiguration_domainConfigurationName - The name of the domain configuration.

Request Lenses

describeDomainConfiguration_domainConfigurationName :: Lens' DescribeDomainConfiguration Text

The name of the domain configuration.

Destructuring the Response

data DescribeDomainConfigurationResponse

See: newDescribeDomainConfigurationResponse smart constructor.

Constructors

DescribeDomainConfigurationResponse'

Fields

domainConfigurationStatus :: Maybe DomainConfigurationStatus

A Boolean value that specifies the current state of the domain configuration.

authorizerConfig :: Maybe AuthorizerConfig

An object that specifies the authorization service for a domain.

serverCertificates :: Maybe [ServerCertificateSummary]

A list containing summary information about the server certificate included in the domain configuration.

domainConfigurationArn :: Maybe Text

The ARN of the domain configuration.

domainName :: Maybe Text

The name of the domain.

domainConfigurationName :: Maybe Text

The name of the domain configuration.

lastStatusChangeDate :: Maybe POSIX

The date and time the domain configuration's status was last changed.

domainType :: Maybe DomainType

The type of the domain.

serviceType :: Maybe ServiceType

The type of service delivered by the endpoint.

httpStatus :: Int

The response's http status code.

newDescribeDomainConfigurationResponse

Arguments

:: Int
$sel:httpStatus:DescribeDomainConfigurationResponse'

-> DescribeDomainConfigurationResponse

Create a value of DescribeDomainConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainConfigurationStatus:DescribeDomainConfigurationResponse', describeDomainConfigurationResponse_domainConfigurationStatus - A Boolean value that specifies the current state of the domain configuration.

$sel:authorizerConfig:DescribeDomainConfigurationResponse', describeDomainConfigurationResponse_authorizerConfig - An object that specifies the authorization service for a domain.

$sel:serverCertificates:DescribeDomainConfigurationResponse', describeDomainConfigurationResponse_serverCertificates - A list containing summary information about the server certificate included in the domain configuration.

$sel:domainConfigurationArn:DescribeDomainConfigurationResponse', describeDomainConfigurationResponse_domainConfigurationArn - The ARN of the domain configuration.

$sel:domainName:DescribeDomainConfigurationResponse', describeDomainConfigurationResponse_domainName - The name of the domain.

$sel:domainConfigurationName:DescribeDomainConfiguration', describeDomainConfigurationResponse_domainConfigurationName - The name of the domain configuration.

$sel:lastStatusChangeDate:DescribeDomainConfigurationResponse', describeDomainConfigurationResponse_lastStatusChangeDate - The date and time the domain configuration's status was last changed.

$sel:domainType:DescribeDomainConfigurationResponse', describeDomainConfigurationResponse_domainType - The type of the domain.

$sel:serviceType:DescribeDomainConfigurationResponse', describeDomainConfigurationResponse_serviceType - The type of service delivered by the endpoint.

$sel:httpStatus:DescribeDomainConfigurationResponse', describeDomainConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

describeDomainConfigurationResponse_domainConfigurationStatus :: Lens' DescribeDomainConfigurationResponse (Maybe DomainConfigurationStatus)

A Boolean value that specifies the current state of the domain configuration.

describeDomainConfigurationResponse_authorizerConfig :: Lens' DescribeDomainConfigurationResponse (Maybe AuthorizerConfig)

An object that specifies the authorization service for a domain.

describeDomainConfigurationResponse_serverCertificates :: Lens' DescribeDomainConfigurationResponse (Maybe [ServerCertificateSummary])

A list containing summary information about the server certificate included in the domain configuration.

describeDomainConfigurationResponse_domainConfigurationArn :: Lens' DescribeDomainConfigurationResponse (Maybe Text)

The ARN of the domain configuration.

describeDomainConfigurationResponse_domainName :: Lens' DescribeDomainConfigurationResponse (Maybe Text)

The name of the domain.

describeDomainConfigurationResponse_domainConfigurationName :: Lens' DescribeDomainConfigurationResponse (Maybe Text)

The name of the domain configuration.

describeDomainConfigurationResponse_lastStatusChangeDate :: Lens' DescribeDomainConfigurationResponse (Maybe UTCTime)

The date and time the domain configuration's status was last changed.

describeDomainConfigurationResponse_domainType :: Lens' DescribeDomainConfigurationResponse (Maybe DomainType)

The type of the domain.

describeDomainConfigurationResponse_serviceType :: Lens' DescribeDomainConfigurationResponse (Maybe ServiceType)

The type of service delivered by the endpoint.

describeDomainConfigurationResponse_httpStatus :: Lens' DescribeDomainConfigurationResponse Int

The response's http status code.