Network.AWS.IoT.DescribeFleetMetric

Description

Gets information about the specified fleet metric.

Requires permission to access the DescribeFleetMetric action.

Creating a Request

data DescribeFleetMetric

See: newDescribeFleetMetric smart constructor.

Constructors

DescribeFleetMetric'

Fields

metricName :: Text

The name of the fleet metric to describe.

newDescribeFleetMetric

Arguments

:: Text
$sel:metricName:DescribeFleetMetric'

-> DescribeFleetMetric

Create a value of DescribeFleetMetric with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricName:DescribeFleetMetric', describeFleetMetric_metricName - The name of the fleet metric to describe.

Request Lenses

describeFleetMetric_metricName :: Lens' DescribeFleetMetric Text

The name of the fleet metric to describe.

Destructuring the Response

data DescribeFleetMetricResponse

See: newDescribeFleetMetricResponse smart constructor.

Constructors

DescribeFleetMetricResponse'

Fields

aggregationType :: Maybe AggregationType

The type of the aggregation query.

lastModifiedDate :: Maybe POSIX

The date when the fleet metric is last modified.

period :: Maybe Natural

The time in seconds between fleet metric emissions. Range [60(1 min), 86400(1 day)] and must be multiple of 60.

queryVersion :: Maybe Text

The query version.

metricName :: Maybe Text

The name of the fleet metric to describe.

aggregationField :: Maybe Text

The field to aggregate.

queryString :: Maybe Text

The search query string.

version :: Maybe Integer

The version of the fleet metric.

creationDate :: Maybe POSIX

The date when the fleet metric is created.

description :: Maybe Text

The fleet metric description.

unit :: Maybe FleetMetricUnit

Used to support unit transformation such as milliseconds to seconds. The unit must be supported by CW metric.

indexName :: Maybe Text

The name of the index to search.

metricArn :: Maybe Text

The ARN of the fleet metric to describe.

httpStatus :: Int

The response's http status code.

newDescribeFleetMetricResponse

Arguments

:: Int
$sel:httpStatus:DescribeFleetMetricResponse'

-> DescribeFleetMetricResponse

Create a value of DescribeFleetMetricResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aggregationType:DescribeFleetMetricResponse', describeFleetMetricResponse_aggregationType - The type of the aggregation query.

$sel:lastModifiedDate:DescribeFleetMetricResponse', describeFleetMetricResponse_lastModifiedDate - The date when the fleet metric is last modified.

$sel:period:DescribeFleetMetricResponse', describeFleetMetricResponse_period - The time in seconds between fleet metric emissions. Range [60(1 min), 86400(1 day)] and must be multiple of 60.

$sel:queryVersion:DescribeFleetMetricResponse', describeFleetMetricResponse_queryVersion - The query version.

$sel:metricName:DescribeFleetMetric', describeFleetMetricResponse_metricName - The name of the fleet metric to describe.

$sel:aggregationField:DescribeFleetMetricResponse', describeFleetMetricResponse_aggregationField - The field to aggregate.

$sel:queryString:DescribeFleetMetricResponse', describeFleetMetricResponse_queryString - The search query string.

$sel:version:DescribeFleetMetricResponse', describeFleetMetricResponse_version - The version of the fleet metric.

$sel:creationDate:DescribeFleetMetricResponse', describeFleetMetricResponse_creationDate - The date when the fleet metric is created.

$sel:description:DescribeFleetMetricResponse', describeFleetMetricResponse_description - The fleet metric description.

$sel:unit:DescribeFleetMetricResponse', describeFleetMetricResponse_unit - Used to support unit transformation such as milliseconds to seconds. The unit must be supported by CW metric.

$sel:indexName:DescribeFleetMetricResponse', describeFleetMetricResponse_indexName - The name of the index to search.

$sel:metricArn:DescribeFleetMetricResponse', describeFleetMetricResponse_metricArn - The ARN of the fleet metric to describe.

$sel:httpStatus:DescribeFleetMetricResponse', describeFleetMetricResponse_httpStatus - The response's http status code.

Response Lenses

describeFleetMetricResponse_aggregationType :: Lens' DescribeFleetMetricResponse (Maybe AggregationType)

The type of the aggregation query.

describeFleetMetricResponse_lastModifiedDate :: Lens' DescribeFleetMetricResponse (Maybe UTCTime)

The date when the fleet metric is last modified.

describeFleetMetricResponse_period :: Lens' DescribeFleetMetricResponse (Maybe Natural)

The time in seconds between fleet metric emissions. Range [60(1 min), 86400(1 day)] and must be multiple of 60.

describeFleetMetricResponse_queryVersion :: Lens' DescribeFleetMetricResponse (Maybe Text)

The query version.

describeFleetMetricResponse_metricName :: Lens' DescribeFleetMetricResponse (Maybe Text)

The name of the fleet metric to describe.

describeFleetMetricResponse_aggregationField :: Lens' DescribeFleetMetricResponse (Maybe Text)

The field to aggregate.

describeFleetMetricResponse_queryString :: Lens' DescribeFleetMetricResponse (Maybe Text)

The search query string.

describeFleetMetricResponse_version :: Lens' DescribeFleetMetricResponse (Maybe Integer)

The version of the fleet metric.

describeFleetMetricResponse_creationDate :: Lens' DescribeFleetMetricResponse (Maybe UTCTime)

The date when the fleet metric is created.

describeFleetMetricResponse_description :: Lens' DescribeFleetMetricResponse (Maybe Text)

The fleet metric description.

describeFleetMetricResponse_unit :: Lens' DescribeFleetMetricResponse (Maybe FleetMetricUnit)

Used to support unit transformation such as milliseconds to seconds. The unit must be supported by CW metric.

describeFleetMetricResponse_indexName :: Lens' DescribeFleetMetricResponse (Maybe Text)

The name of the index to search.

describeFleetMetricResponse_metricArn :: Lens' DescribeFleetMetricResponse (Maybe Text)

The ARN of the fleet metric to describe.

describeFleetMetricResponse_httpStatus :: Lens' DescribeFleetMetricResponse Int

The response's http status code.