Network.AWS.IoT.DescribeJobTemplate

Description

Returns information about a job template.

Creating a Request

data DescribeJobTemplate

See: newDescribeJobTemplate smart constructor.

Constructors

DescribeJobTemplate'

Fields

jobTemplateId :: Text

The unique identifier of the job template.

newDescribeJobTemplate

Arguments

:: Text
$sel:jobTemplateId:DescribeJobTemplate'

-> DescribeJobTemplate

Create a value of DescribeJobTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobTemplateId:DescribeJobTemplate', describeJobTemplate_jobTemplateId - The unique identifier of the job template.

Request Lenses

describeJobTemplate_jobTemplateId :: Lens' DescribeJobTemplate Text

The unique identifier of the job template.

Destructuring the Response

data DescribeJobTemplateResponse

See: newDescribeJobTemplateResponse smart constructor.

Constructors

DescribeJobTemplateResponse'

Fields

jobExecutionsRolloutConfig :: Maybe JobExecutionsRolloutConfig

timeoutConfig :: Maybe TimeoutConfig

createdAt :: Maybe POSIX

The time, in seconds since the epoch, when the job template was created.

jobTemplateId :: Maybe Text

The unique identifier of the job template.

documentSource :: Maybe Text

An S3 link to the job document.

document :: Maybe Text

The job document.

presignedUrlConfig :: Maybe PresignedUrlConfig

description :: Maybe Text

A description of the job template.

abortConfig :: Maybe AbortConfig

jobTemplateArn :: Maybe Text

The ARN of the job template.

httpStatus :: Int

The response's http status code.

newDescribeJobTemplateResponse

Arguments

:: Int
$sel:httpStatus:DescribeJobTemplateResponse'

-> DescribeJobTemplateResponse

Create a value of DescribeJobTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobExecutionsRolloutConfig:DescribeJobTemplateResponse', describeJobTemplateResponse_jobExecutionsRolloutConfig - Undocumented member.

$sel:timeoutConfig:DescribeJobTemplateResponse', describeJobTemplateResponse_timeoutConfig - Undocumented member.

$sel:createdAt:DescribeJobTemplateResponse', describeJobTemplateResponse_createdAt - The time, in seconds since the epoch, when the job template was created.

$sel:jobTemplateId:DescribeJobTemplate', describeJobTemplateResponse_jobTemplateId - The unique identifier of the job template.

$sel:documentSource:DescribeJobTemplateResponse', describeJobTemplateResponse_documentSource - An S3 link to the job document.

$sel:document:DescribeJobTemplateResponse', describeJobTemplateResponse_document - The job document.

$sel:presignedUrlConfig:DescribeJobTemplateResponse', describeJobTemplateResponse_presignedUrlConfig - Undocumented member.

$sel:description:DescribeJobTemplateResponse', describeJobTemplateResponse_description - A description of the job template.

$sel:abortConfig:DescribeJobTemplateResponse', describeJobTemplateResponse_abortConfig - Undocumented member.

$sel:jobTemplateArn:DescribeJobTemplateResponse', describeJobTemplateResponse_jobTemplateArn - The ARN of the job template.

$sel:httpStatus:DescribeJobTemplateResponse', describeJobTemplateResponse_httpStatus - The response's http status code.

Response Lenses

describeJobTemplateResponse_jobExecutionsRolloutConfig :: Lens' DescribeJobTemplateResponse (Maybe JobExecutionsRolloutConfig)

Undocumented member.

describeJobTemplateResponse_timeoutConfig :: Lens' DescribeJobTemplateResponse (Maybe TimeoutConfig)

Undocumented member.

describeJobTemplateResponse_createdAt :: Lens' DescribeJobTemplateResponse (Maybe UTCTime)

The time, in seconds since the epoch, when the job template was created.

describeJobTemplateResponse_jobTemplateId :: Lens' DescribeJobTemplateResponse (Maybe Text)

The unique identifier of the job template.

describeJobTemplateResponse_documentSource :: Lens' DescribeJobTemplateResponse (Maybe Text)

An S3 link to the job document.

describeJobTemplateResponse_document :: Lens' DescribeJobTemplateResponse (Maybe Text)

The job document.

describeJobTemplateResponse_presignedUrlConfig :: Lens' DescribeJobTemplateResponse (Maybe PresignedUrlConfig)

Undocumented member.

describeJobTemplateResponse_description :: Lens' DescribeJobTemplateResponse (Maybe Text)

A description of the job template.

describeJobTemplateResponse_abortConfig :: Lens' DescribeJobTemplateResponse (Maybe AbortConfig)

Undocumented member.

describeJobTemplateResponse_jobTemplateArn :: Lens' DescribeJobTemplateResponse (Maybe Text)

The ARN of the job template.

describeJobTemplateResponse_httpStatus :: Lens' DescribeJobTemplateResponse Int

The response's http status code.