Network.AWS.IoT.DescribeProvisioningTemplate

Description

Returns information about a fleet provisioning template.

Requires permission to access the DescribeProvisioningTemplate action.

Creating a Request

data DescribeProvisioningTemplate

See: newDescribeProvisioningTemplate smart constructor.

Constructors

DescribeProvisioningTemplate'

Fields

templateName :: Text

The name of the fleet provisioning template.

newDescribeProvisioningTemplate

Arguments

:: Text
$sel:templateName:DescribeProvisioningTemplate'

-> DescribeProvisioningTemplate

Create a value of DescribeProvisioningTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:DescribeProvisioningTemplate', describeProvisioningTemplate_templateName - The name of the fleet provisioning template.

Request Lenses

describeProvisioningTemplate_templateName :: Lens' DescribeProvisioningTemplate Text

The name of the fleet provisioning template.

Destructuring the Response

data DescribeProvisioningTemplateResponse

See: newDescribeProvisioningTemplateResponse smart constructor.

Constructors

DescribeProvisioningTemplateResponse'

Fields

lastModifiedDate :: Maybe POSIX

The date when the fleet provisioning template was last modified.

templateName :: Maybe Text

The name of the fleet provisioning template.

preProvisioningHook :: Maybe ProvisioningHook

Gets information about a pre-provisioned hook.

enabled :: Maybe Bool

True if the fleet provisioning template is enabled, otherwise false.

provisioningRoleArn :: Maybe Text

The ARN of the role associated with the provisioning template. This IoT role grants permission to provision a device.

defaultVersionId :: Maybe Int

The default fleet template version ID.

creationDate :: Maybe POSIX

The date when the fleet provisioning template was created.

templateArn :: Maybe Text

The ARN of the fleet provisioning template.

templateBody :: Maybe Text

The JSON formatted contents of the fleet provisioning template.

description :: Maybe Text

The description of the fleet provisioning template.

httpStatus :: Int

The response's http status code.

newDescribeProvisioningTemplateResponse

Arguments

:: Int
$sel:httpStatus:DescribeProvisioningTemplateResponse'

-> DescribeProvisioningTemplateResponse

Create a value of DescribeProvisioningTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:DescribeProvisioningTemplateResponse', describeProvisioningTemplateResponse_lastModifiedDate - The date when the fleet provisioning template was last modified.

$sel:templateName:DescribeProvisioningTemplate', describeProvisioningTemplateResponse_templateName - The name of the fleet provisioning template.

$sel:preProvisioningHook:DescribeProvisioningTemplateResponse', describeProvisioningTemplateResponse_preProvisioningHook - Gets information about a pre-provisioned hook.

$sel:enabled:DescribeProvisioningTemplateResponse', describeProvisioningTemplateResponse_enabled - True if the fleet provisioning template is enabled, otherwise false.

$sel:provisioningRoleArn:DescribeProvisioningTemplateResponse', describeProvisioningTemplateResponse_provisioningRoleArn - The ARN of the role associated with the provisioning template. This IoT role grants permission to provision a device.

$sel:defaultVersionId:DescribeProvisioningTemplateResponse', describeProvisioningTemplateResponse_defaultVersionId - The default fleet template version ID.

$sel:creationDate:DescribeProvisioningTemplateResponse', describeProvisioningTemplateResponse_creationDate - The date when the fleet provisioning template was created.

$sel:templateArn:DescribeProvisioningTemplateResponse', describeProvisioningTemplateResponse_templateArn - The ARN of the fleet provisioning template.

$sel:templateBody:DescribeProvisioningTemplateResponse', describeProvisioningTemplateResponse_templateBody - The JSON formatted contents of the fleet provisioning template.

$sel:description:DescribeProvisioningTemplateResponse', describeProvisioningTemplateResponse_description - The description of the fleet provisioning template.

$sel:httpStatus:DescribeProvisioningTemplateResponse', describeProvisioningTemplateResponse_httpStatus - The response's http status code.

Response Lenses

describeProvisioningTemplateResponse_lastModifiedDate :: Lens' DescribeProvisioningTemplateResponse (Maybe UTCTime)

The date when the fleet provisioning template was last modified.

describeProvisioningTemplateResponse_templateName :: Lens' DescribeProvisioningTemplateResponse (Maybe Text)

The name of the fleet provisioning template.

describeProvisioningTemplateResponse_preProvisioningHook :: Lens' DescribeProvisioningTemplateResponse (Maybe ProvisioningHook)

Gets information about a pre-provisioned hook.

describeProvisioningTemplateResponse_enabled :: Lens' DescribeProvisioningTemplateResponse (Maybe Bool)

True if the fleet provisioning template is enabled, otherwise false.

describeProvisioningTemplateResponse_provisioningRoleArn :: Lens' DescribeProvisioningTemplateResponse (Maybe Text)

The ARN of the role associated with the provisioning template. This IoT role grants permission to provision a device.

describeProvisioningTemplateResponse_defaultVersionId :: Lens' DescribeProvisioningTemplateResponse (Maybe Int)

The default fleet template version ID.

describeProvisioningTemplateResponse_creationDate :: Lens' DescribeProvisioningTemplateResponse (Maybe UTCTime)

The date when the fleet provisioning template was created.

describeProvisioningTemplateResponse_templateArn :: Lens' DescribeProvisioningTemplateResponse (Maybe Text)

The ARN of the fleet provisioning template.

describeProvisioningTemplateResponse_templateBody :: Lens' DescribeProvisioningTemplateResponse (Maybe Text)

The JSON formatted contents of the fleet provisioning template.

describeProvisioningTemplateResponse_description :: Lens' DescribeProvisioningTemplateResponse (Maybe Text)

The description of the fleet provisioning template.

describeProvisioningTemplateResponse_httpStatus :: Lens' DescribeProvisioningTemplateResponse Int

The response's http status code.