Network.AWS.IoT.DescribeProvisioningTemplateVersion

Description

Returns information about a fleet provisioning template version.

Requires permission to access the DescribeProvisioningTemplateVersion action.

Creating a Request

data DescribeProvisioningTemplateVersion

See: newDescribeProvisioningTemplateVersion smart constructor.

Constructors

DescribeProvisioningTemplateVersion'

Fields

templateName :: Text

The template name.

versionId :: Int

The fleet provisioning template version ID.

newDescribeProvisioningTemplateVersion

Arguments

:: Text
$sel:templateName:DescribeProvisioningTemplateVersion'

-> Int
$sel:versionId:DescribeProvisioningTemplateVersion'

-> DescribeProvisioningTemplateVersion

Create a value of DescribeProvisioningTemplateVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:DescribeProvisioningTemplateVersion', describeProvisioningTemplateVersion_templateName - The template name.

$sel:versionId:DescribeProvisioningTemplateVersion', describeProvisioningTemplateVersion_versionId - The fleet provisioning template version ID.

Request Lenses

describeProvisioningTemplateVersion_templateName :: Lens' DescribeProvisioningTemplateVersion Text

The template name.

describeProvisioningTemplateVersion_versionId :: Lens' DescribeProvisioningTemplateVersion Int

The fleet provisioning template version ID.

Destructuring the Response

data DescribeProvisioningTemplateVersionResponse

See: newDescribeProvisioningTemplateVersionResponse smart constructor.

Constructors

DescribeProvisioningTemplateVersionResponse'

Fields

creationDate :: Maybe POSIX

The date when the fleet provisioning template version was created.

versionId :: Maybe Int

The fleet provisioning template version ID.

isDefaultVersion :: Maybe Bool

True if the fleet provisioning template version is the default version.

templateBody :: Maybe Text

The JSON formatted contents of the fleet provisioning template version.

httpStatus :: Int

The response's http status code.

newDescribeProvisioningTemplateVersionResponse

Arguments

:: Int
$sel:httpStatus:DescribeProvisioningTemplateVersionResponse'

-> DescribeProvisioningTemplateVersionResponse

Create a value of DescribeProvisioningTemplateVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationDate:DescribeProvisioningTemplateVersionResponse', describeProvisioningTemplateVersionResponse_creationDate - The date when the fleet provisioning template version was created.

$sel:versionId:DescribeProvisioningTemplateVersion', describeProvisioningTemplateVersionResponse_versionId - The fleet provisioning template version ID.

$sel:isDefaultVersion:DescribeProvisioningTemplateVersionResponse', describeProvisioningTemplateVersionResponse_isDefaultVersion - True if the fleet provisioning template version is the default version.

$sel:templateBody:DescribeProvisioningTemplateVersionResponse', describeProvisioningTemplateVersionResponse_templateBody - The JSON formatted contents of the fleet provisioning template version.

$sel:httpStatus:DescribeProvisioningTemplateVersionResponse', describeProvisioningTemplateVersionResponse_httpStatus - The response's http status code.

Response Lenses

describeProvisioningTemplateVersionResponse_creationDate :: Lens' DescribeProvisioningTemplateVersionResponse (Maybe UTCTime)

The date when the fleet provisioning template version was created.

describeProvisioningTemplateVersionResponse_versionId :: Lens' DescribeProvisioningTemplateVersionResponse (Maybe Int)

The fleet provisioning template version ID.

describeProvisioningTemplateVersionResponse_isDefaultVersion :: Lens' DescribeProvisioningTemplateVersionResponse (Maybe Bool)

True if the fleet provisioning template version is the default version.

describeProvisioningTemplateVersionResponse_templateBody :: Lens' DescribeProvisioningTemplateVersionResponse (Maybe Text)

The JSON formatted contents of the fleet provisioning template version.

describeProvisioningTemplateVersionResponse_httpStatus :: Lens' DescribeProvisioningTemplateVersionResponse Int

The response's http status code.