Network.AWS.IoT.DescribeThing

Description

Gets information about the specified thing.

Requires permission to access the DescribeThing action.

Creating a Request

data DescribeThing

The input for the DescribeThing operation.

See: newDescribeThing smart constructor.

Constructors

DescribeThing'

Fields

thingName :: Text

The name of the thing.

newDescribeThing

Arguments

:: Text
$sel:thingName:DescribeThing'

-> DescribeThing

Create a value of DescribeThing with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thingName:DescribeThing', describeThing_thingName - The name of the thing.

Request Lenses

describeThing_thingName :: Lens' DescribeThing Text

The name of the thing.

Destructuring the Response

data DescribeThingResponse

The output from the DescribeThing operation.

See: newDescribeThingResponse smart constructor.

Constructors

DescribeThingResponse'

Fields

defaultClientId :: Maybe Text

The default MQTT client ID. For a typical device, the thing name is also used as the default MQTT client ID. Although we don’t require a mapping between a thing's registry name and its use of MQTT client IDs, certificates, or shadow state, we recommend that you choose a thing name and use it as the MQTT client ID for the registry and the Device Shadow service.

This lets you better organize your IoT fleet without removing the flexibility of the underlying device certificate model or shadows.

thingTypeName :: Maybe Text

The thing type name.

thingArn :: Maybe Text

The ARN of the thing to describe.

attributes :: Maybe (HashMap Text Text)

The thing attributes.

version :: Maybe Integer

The current version of the thing record in the registry.

To avoid unintentional changes to the information in the registry, you can pass the version information in the expectedVersion parameter of the UpdateThing and DeleteThing calls.

thingName :: Maybe Text

The name of the thing.

billingGroupName :: Maybe Text

The name of the billing group the thing belongs to.

thingId :: Maybe Text

The ID of the thing to describe.

httpStatus :: Int

The response's http status code.

newDescribeThingResponse

Arguments

:: Int
$sel:httpStatus:DescribeThingResponse'

-> DescribeThingResponse

Create a value of DescribeThingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defaultClientId:DescribeThingResponse', describeThingResponse_defaultClientId - The default MQTT client ID. For a typical device, the thing name is also used as the default MQTT client ID. Although we don’t require a mapping between a thing's registry name and its use of MQTT client IDs, certificates, or shadow state, we recommend that you choose a thing name and use it as the MQTT client ID for the registry and the Device Shadow service.

This lets you better organize your IoT fleet without removing the flexibility of the underlying device certificate model or shadows.

$sel:thingTypeName:DescribeThingResponse', describeThingResponse_thingTypeName - The thing type name.

$sel:thingArn:DescribeThingResponse', describeThingResponse_thingArn - The ARN of the thing to describe.

$sel:attributes:DescribeThingResponse', describeThingResponse_attributes - The thing attributes.

$sel:version:DescribeThingResponse', describeThingResponse_version - The current version of the thing record in the registry.

To avoid unintentional changes to the information in the registry, you can pass the version information in the expectedVersion parameter of the UpdateThing and DeleteThing calls.

$sel:thingName:DescribeThing', describeThingResponse_thingName - The name of the thing.

$sel:billingGroupName:DescribeThingResponse', describeThingResponse_billingGroupName - The name of the billing group the thing belongs to.

$sel:thingId:DescribeThingResponse', describeThingResponse_thingId - The ID of the thing to describe.

$sel:httpStatus:DescribeThingResponse', describeThingResponse_httpStatus - The response's http status code.

Response Lenses

describeThingResponse_defaultClientId :: Lens' DescribeThingResponse (Maybe Text)

The default MQTT client ID. For a typical device, the thing name is also used as the default MQTT client ID. Although we don’t require a mapping between a thing's registry name and its use of MQTT client IDs, certificates, or shadow state, we recommend that you choose a thing name and use it as the MQTT client ID for the registry and the Device Shadow service.

This lets you better organize your IoT fleet without removing the flexibility of the underlying device certificate model or shadows.

describeThingResponse_thingTypeName :: Lens' DescribeThingResponse (Maybe Text)

The thing type name.

describeThingResponse_thingArn :: Lens' DescribeThingResponse (Maybe Text)

The ARN of the thing to describe.

describeThingResponse_attributes :: Lens' DescribeThingResponse (Maybe (HashMap Text Text))

The thing attributes.

describeThingResponse_version :: Lens' DescribeThingResponse (Maybe Integer)

The current version of the thing record in the registry.

To avoid unintentional changes to the information in the registry, you can pass the version information in the expectedVersion parameter of the UpdateThing and DeleteThing calls.

describeThingResponse_thingName :: Lens' DescribeThingResponse (Maybe Text)

The name of the thing.

describeThingResponse_billingGroupName :: Lens' DescribeThingResponse (Maybe Text)

The name of the billing group the thing belongs to.

describeThingResponse_thingId :: Lens' DescribeThingResponse (Maybe Text)

The ID of the thing to describe.

describeThingResponse_httpStatus :: Lens' DescribeThingResponse Int

The response's http status code.