Network.AWS.IoT.DescribeThingGroup

Description

Describe a thing group.

Requires permission to access the DescribeThingGroup action.

Creating a Request

data DescribeThingGroup

See: newDescribeThingGroup smart constructor.

Constructors

DescribeThingGroup'

Fields

thingGroupName :: Text

The name of the thing group.

newDescribeThingGroup

Arguments

:: Text
$sel:thingGroupName:DescribeThingGroup'

-> DescribeThingGroup

Create a value of DescribeThingGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thingGroupName:DescribeThingGroup', describeThingGroup_thingGroupName - The name of the thing group.

Request Lenses

describeThingGroup_thingGroupName :: Lens' DescribeThingGroup Text

The name of the thing group.

Destructuring the Response

data DescribeThingGroupResponse

See: newDescribeThingGroupResponse smart constructor.

Constructors

DescribeThingGroupResponse'

Fields

queryString :: Maybe Text

The dynamic thing group search query string.

status :: Maybe DynamicGroupStatus

The dynamic thing group status.

indexName :: Maybe Text

The dynamic thing group index name.

thingGroupArn :: Maybe Text

The thing group ARN.

queryVersion :: Maybe Text

The dynamic thing group query version.

version :: Maybe Integer

The version of the thing group.

thingGroupName :: Maybe Text

The name of the thing group.

thingGroupId :: Maybe Text

The thing group ID.

thingGroupMetadata :: Maybe ThingGroupMetadata

Thing group metadata.

thingGroupProperties :: Maybe ThingGroupProperties

The thing group properties.

httpStatus :: Int

The response's http status code.

newDescribeThingGroupResponse

Arguments

:: Int
$sel:httpStatus:DescribeThingGroupResponse'

-> DescribeThingGroupResponse

Create a value of DescribeThingGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queryString:DescribeThingGroupResponse', describeThingGroupResponse_queryString - The dynamic thing group search query string.

$sel:status:DescribeThingGroupResponse', describeThingGroupResponse_status - The dynamic thing group status.

$sel:indexName:DescribeThingGroupResponse', describeThingGroupResponse_indexName - The dynamic thing group index name.

$sel:thingGroupArn:DescribeThingGroupResponse', describeThingGroupResponse_thingGroupArn - The thing group ARN.

$sel:queryVersion:DescribeThingGroupResponse', describeThingGroupResponse_queryVersion - The dynamic thing group query version.

$sel:version:DescribeThingGroupResponse', describeThingGroupResponse_version - The version of the thing group.

$sel:thingGroupName:DescribeThingGroup', describeThingGroupResponse_thingGroupName - The name of the thing group.

$sel:thingGroupId:DescribeThingGroupResponse', describeThingGroupResponse_thingGroupId - The thing group ID.

$sel:thingGroupMetadata:DescribeThingGroupResponse', describeThingGroupResponse_thingGroupMetadata - Thing group metadata.

$sel:thingGroupProperties:DescribeThingGroupResponse', describeThingGroupResponse_thingGroupProperties - The thing group properties.

$sel:httpStatus:DescribeThingGroupResponse', describeThingGroupResponse_httpStatus - The response's http status code.

Response Lenses

describeThingGroupResponse_queryString :: Lens' DescribeThingGroupResponse (Maybe Text)

The dynamic thing group search query string.

describeThingGroupResponse_status :: Lens' DescribeThingGroupResponse (Maybe DynamicGroupStatus)

The dynamic thing group status.

describeThingGroupResponse_indexName :: Lens' DescribeThingGroupResponse (Maybe Text)

The dynamic thing group index name.

describeThingGroupResponse_thingGroupArn :: Lens' DescribeThingGroupResponse (Maybe Text)

The thing group ARN.

describeThingGroupResponse_queryVersion :: Lens' DescribeThingGroupResponse (Maybe Text)

The dynamic thing group query version.

describeThingGroupResponse_version :: Lens' DescribeThingGroupResponse (Maybe Integer)

The version of the thing group.

describeThingGroupResponse_thingGroupName :: Lens' DescribeThingGroupResponse (Maybe Text)

The name of the thing group.

describeThingGroupResponse_thingGroupId :: Lens' DescribeThingGroupResponse (Maybe Text)

The thing group ID.

describeThingGroupResponse_thingGroupMetadata :: Lens' DescribeThingGroupResponse (Maybe ThingGroupMetadata)

Thing group metadata.

describeThingGroupResponse_thingGroupProperties :: Lens' DescribeThingGroupResponse (Maybe ThingGroupProperties)

The thing group properties.

describeThingGroupResponse_httpStatus :: Lens' DescribeThingGroupResponse Int

The response's http status code.