Network.AWS.IoT.DescribeThingType

Description

Gets information about the specified thing type.

Requires permission to access the DescribeThingType action.

Creating a Request

data DescribeThingType

The input for the DescribeThingType operation.

See: newDescribeThingType smart constructor.

Constructors

DescribeThingType'

Fields

thingTypeName :: Text

The name of the thing type.

newDescribeThingType

Arguments

:: Text
$sel:thingTypeName:DescribeThingType'

-> DescribeThingType

Create a value of DescribeThingType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thingTypeName:DescribeThingType', describeThingType_thingTypeName - The name of the thing type.

Request Lenses

describeThingType_thingTypeName :: Lens' DescribeThingType Text

The name of the thing type.

Destructuring the Response

data DescribeThingTypeResponse

The output for the DescribeThingType operation.

See: newDescribeThingTypeResponse smart constructor.

Constructors

DescribeThingTypeResponse'

Fields

thingTypeProperties :: Maybe ThingTypeProperties

The ThingTypeProperties contains information about the thing type including description, and a list of searchable thing attribute names.

thingTypeMetadata :: Maybe ThingTypeMetadata

The ThingTypeMetadata contains additional information about the thing type including: creation date and time, a value indicating whether the thing type is deprecated, and a date and time when it was deprecated.

thingTypeId :: Maybe Text

The thing type ID.

thingTypeArn :: Maybe Text

The thing type ARN.

thingTypeName :: Maybe Text

The name of the thing type.

httpStatus :: Int

The response's http status code.

newDescribeThingTypeResponse

Arguments

:: Int
$sel:httpStatus:DescribeThingTypeResponse'

-> DescribeThingTypeResponse

Create a value of DescribeThingTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thingTypeProperties:DescribeThingTypeResponse', describeThingTypeResponse_thingTypeProperties - The ThingTypeProperties contains information about the thing type including description, and a list of searchable thing attribute names.

$sel:thingTypeMetadata:DescribeThingTypeResponse', describeThingTypeResponse_thingTypeMetadata - The ThingTypeMetadata contains additional information about the thing type including: creation date and time, a value indicating whether the thing type is deprecated, and a date and time when it was deprecated.

$sel:thingTypeId:DescribeThingTypeResponse', describeThingTypeResponse_thingTypeId - The thing type ID.

$sel:thingTypeArn:DescribeThingTypeResponse', describeThingTypeResponse_thingTypeArn - The thing type ARN.

$sel:thingTypeName:DescribeThingType', describeThingTypeResponse_thingTypeName - The name of the thing type.

$sel:httpStatus:DescribeThingTypeResponse', describeThingTypeResponse_httpStatus - The response's http status code.

Response Lenses

describeThingTypeResponse_thingTypeProperties :: Lens' DescribeThingTypeResponse (Maybe ThingTypeProperties)

The ThingTypeProperties contains information about the thing type including description, and a list of searchable thing attribute names.

describeThingTypeResponse_thingTypeMetadata :: Lens' DescribeThingTypeResponse (Maybe ThingTypeMetadata)

The ThingTypeMetadata contains additional information about the thing type including: creation date and time, a value indicating whether the thing type is deprecated, and a date and time when it was deprecated.

describeThingTypeResponse_thingTypeId :: Lens' DescribeThingTypeResponse (Maybe Text)

The thing type ID.

describeThingTypeResponse_thingTypeArn :: Lens' DescribeThingTypeResponse (Maybe Text)

The thing type ARN.

describeThingTypeResponse_thingTypeName :: Lens' DescribeThingTypeResponse (Maybe Text)

The name of the thing type.

describeThingTypeResponse_httpStatus :: Lens' DescribeThingTypeResponse Int

The response's http status code.