Network.AWS.IoT.DetachThingPrincipal

Description

Detaches the specified principal from the specified thing. A principal can be X.509 certificates, IAM users, groups, and roles, Amazon Cognito identities or federated identities.

This call is asynchronous. It might take several seconds for the detachment to propagate.

Requires permission to access the DetachThingPrincipal action.

Creating a Request

data DetachThingPrincipal

The input for the DetachThingPrincipal operation.

See: newDetachThingPrincipal smart constructor.

Constructors

DetachThingPrincipal'

Fields

thingName :: Text

The name of the thing.

principal :: Text

If the principal is a certificate, this value must be ARN of the certificate. If the principal is an Amazon Cognito identity, this value must be the ID of the Amazon Cognito identity.

newDetachThingPrincipal

Arguments

:: Text
$sel:thingName:DetachThingPrincipal'

-> Text
$sel:principal:DetachThingPrincipal'

-> DetachThingPrincipal

Create a value of DetachThingPrincipal with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thingName:DetachThingPrincipal', detachThingPrincipal_thingName - The name of the thing.

$sel:principal:DetachThingPrincipal', detachThingPrincipal_principal - If the principal is a certificate, this value must be ARN of the certificate. If the principal is an Amazon Cognito identity, this value must be the ID of the Amazon Cognito identity.

Request Lenses

detachThingPrincipal_thingName :: Lens' DetachThingPrincipal Text

The name of the thing.

detachThingPrincipal_principal :: Lens' DetachThingPrincipal Text

If the principal is a certificate, this value must be ARN of the certificate. If the principal is an Amazon Cognito identity, this value must be the ID of the Amazon Cognito identity.

Destructuring the Response

data DetachThingPrincipalResponse

The output from the DetachThingPrincipal operation.

See: newDetachThingPrincipalResponse smart constructor.

Constructors

DetachThingPrincipalResponse'

Fields

httpStatus :: Int

The response's http status code.

newDetachThingPrincipalResponse

Arguments

:: Int
$sel:httpStatus:DetachThingPrincipalResponse'

-> DetachThingPrincipalResponse

Create a value of DetachThingPrincipalResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DetachThingPrincipalResponse', detachThingPrincipalResponse_httpStatus - The response's http status code.

Response Lenses

detachThingPrincipalResponse_httpStatus :: Lens' DetachThingPrincipalResponse Int

The response's http status code.