Network.AWS.IoT.GetBucketsAggregation

Description

Aggregates on indexed data with search queries pertaining to particular fields.

Requires permission to access the GetBucketsAggregation action.

Creating a Request

data GetBucketsAggregation

See: newGetBucketsAggregation smart constructor.

Constructors

GetBucketsAggregation'

Fields

queryVersion :: Maybe Text

The version of the query.

indexName :: Maybe Text

The name of the index to search.

queryString :: Text

The search query string.

aggregationField :: Text

The aggregation field.

bucketsAggregationType :: BucketsAggregationType

The basic control of the response shape and the bucket aggregation type to perform.

newGetBucketsAggregation

Arguments

:: Text
$sel:queryString:GetBucketsAggregation'

-> Text
$sel:aggregationField:GetBucketsAggregation'

-> BucketsAggregationType
$sel:bucketsAggregationType:GetBucketsAggregation'

-> GetBucketsAggregation

Create a value of GetBucketsAggregation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queryVersion:GetBucketsAggregation', getBucketsAggregation_queryVersion - The version of the query.

$sel:indexName:GetBucketsAggregation', getBucketsAggregation_indexName - The name of the index to search.

$sel:queryString:GetBucketsAggregation', getBucketsAggregation_queryString - The search query string.

$sel:aggregationField:GetBucketsAggregation', getBucketsAggregation_aggregationField - The aggregation field.

$sel:bucketsAggregationType:GetBucketsAggregation', getBucketsAggregation_bucketsAggregationType - The basic control of the response shape and the bucket aggregation type to perform.

Request Lenses

getBucketsAggregation_queryVersion :: Lens' GetBucketsAggregation (Maybe Text)

The version of the query.

getBucketsAggregation_indexName :: Lens' GetBucketsAggregation (Maybe Text)

The name of the index to search.

getBucketsAggregation_queryString :: Lens' GetBucketsAggregation Text

The search query string.

getBucketsAggregation_aggregationField :: Lens' GetBucketsAggregation Text

The aggregation field.

getBucketsAggregation_bucketsAggregationType :: Lens' GetBucketsAggregation BucketsAggregationType

The basic control of the response shape and the bucket aggregation type to perform.

Destructuring the Response

data GetBucketsAggregationResponse

See: newGetBucketsAggregationResponse smart constructor.

Constructors

GetBucketsAggregationResponse'

Fields

buckets :: Maybe [Bucket]

The main part of the response with a list of buckets. Each bucket contains a keyValue and a count.

keyValue: The aggregation field value counted for the particular bucket.

count: The number of documents that have that value.

totalCount :: Maybe Int

The total number of documents that fit the query string criteria and contain a value for the Aggregation field targeted in the request.

httpStatus :: Int

The response's http status code.

newGetBucketsAggregationResponse

Arguments

:: Int
$sel:httpStatus:GetBucketsAggregationResponse'

-> GetBucketsAggregationResponse

Create a value of GetBucketsAggregationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:buckets:GetBucketsAggregationResponse', getBucketsAggregationResponse_buckets - The main part of the response with a list of buckets. Each bucket contains a keyValue and a count.

keyValue: The aggregation field value counted for the particular bucket.

count: The number of documents that have that value.

$sel:totalCount:GetBucketsAggregationResponse', getBucketsAggregationResponse_totalCount - The total number of documents that fit the query string criteria and contain a value for the Aggregation field targeted in the request.

$sel:httpStatus:GetBucketsAggregationResponse', getBucketsAggregationResponse_httpStatus - The response's http status code.

Response Lenses

getBucketsAggregationResponse_buckets :: Lens' GetBucketsAggregationResponse (Maybe [Bucket])

The main part of the response with a list of buckets. Each bucket contains a keyValue and a count.

keyValue: The aggregation field value counted for the particular bucket.

count: The number of documents that have that value.

getBucketsAggregationResponse_totalCount :: Lens' GetBucketsAggregationResponse (Maybe Int)

The total number of documents that fit the query string criteria and contain a value for the Aggregation field targeted in the request.

getBucketsAggregationResponse_httpStatus :: Lens' GetBucketsAggregationResponse Int

The response's http status code.