Network.AWS.IoT.GetEffectivePolicies

Description

Gets a list of the policies that have an effect on the authorization behavior of the specified device when it connects to the IoT device gateway.

Requires permission to access the GetEffectivePolicies action.

Creating a Request

data GetEffectivePolicies

See: newGetEffectivePolicies smart constructor.

Constructors

GetEffectivePolicies'

Fields

principal :: Maybe Text

The principal. Valid principals are CertificateArn (arn:aws:iot:region:accountId:cert/certificateId), thingGroupArn (arn:aws:iot:region:accountId:thinggroup/groupName) and CognitoId (region:id).

cognitoIdentityPoolId :: Maybe Text

The Cognito identity pool ID.

thingName :: Maybe Text

The thing name.

newGetEffectivePolicies :: GetEffectivePolicies

Create a value of GetEffectivePolicies with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:principal:GetEffectivePolicies', getEffectivePolicies_principal - The principal. Valid principals are CertificateArn (arn:aws:iot:region:accountId:cert/certificateId), thingGroupArn (arn:aws:iot:region:accountId:thinggroup/groupName) and CognitoId (region:id).

$sel:cognitoIdentityPoolId:GetEffectivePolicies', getEffectivePolicies_cognitoIdentityPoolId - The Cognito identity pool ID.

$sel:thingName:GetEffectivePolicies', getEffectivePolicies_thingName - The thing name.

Request Lenses

getEffectivePolicies_principal :: Lens' GetEffectivePolicies (Maybe Text)

The principal. Valid principals are CertificateArn (arn:aws:iot:region:accountId:cert/certificateId), thingGroupArn (arn:aws:iot:region:accountId:thinggroup/groupName) and CognitoId (region:id).

getEffectivePolicies_cognitoIdentityPoolId :: Lens' GetEffectivePolicies (Maybe Text)

The Cognito identity pool ID.

getEffectivePolicies_thingName :: Lens' GetEffectivePolicies (Maybe Text)

The thing name.

Destructuring the Response

data GetEffectivePoliciesResponse

See: newGetEffectivePoliciesResponse smart constructor.

Constructors

GetEffectivePoliciesResponse'

Fields

effectivePolicies :: Maybe [EffectivePolicy]

The effective policies.

httpStatus :: Int

The response's http status code.

newGetEffectivePoliciesResponse

Arguments

:: Int
$sel:httpStatus:GetEffectivePoliciesResponse'

-> GetEffectivePoliciesResponse

Create a value of GetEffectivePoliciesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:effectivePolicies:GetEffectivePoliciesResponse', getEffectivePoliciesResponse_effectivePolicies - The effective policies.

$sel:httpStatus:GetEffectivePoliciesResponse', getEffectivePoliciesResponse_httpStatus - The response's http status code.

Response Lenses

getEffectivePoliciesResponse_effectivePolicies :: Lens' GetEffectivePoliciesResponse (Maybe [EffectivePolicy])

The effective policies.

getEffectivePoliciesResponse_httpStatus :: Lens' GetEffectivePoliciesResponse Int

The response's http status code.