Network.AWS.IoT.GetPercentiles

Description

Groups the aggregated values that match the query into percentile groupings. The default percentile groupings are: 1,5,25,50,75,95,99, although you can specify your own when you call GetPercentiles. This function returns a value for each percentile group specified (or the default percentile groupings). The percentile group "1" contains the aggregated field value that occurs in approximately one percent of the values that match the query. The percentile group "5" contains the aggregated field value that occurs in approximately five percent of the values that match the query, and so on. The result is an approximation, the more values that match the query, the more accurate the percentile values.

Requires permission to access the GetPercentiles action.

Creating a Request

data GetPercentiles

See: newGetPercentiles smart constructor.

Constructors

GetPercentiles'

Fields

percents :: Maybe [Double]

The percentile groups returned.

queryVersion :: Maybe Text

The query version.

aggregationField :: Maybe Text

The field to aggregate.

indexName :: Maybe Text

The name of the index to search.

queryString :: Text

The search query string.

newGetPercentiles

Arguments

:: Text
$sel:queryString:GetPercentiles'

-> GetPercentiles

Create a value of GetPercentiles with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:percents:GetPercentiles', getPercentiles_percents - The percentile groups returned.

$sel:queryVersion:GetPercentiles', getPercentiles_queryVersion - The query version.

$sel:aggregationField:GetPercentiles', getPercentiles_aggregationField - The field to aggregate.

$sel:indexName:GetPercentiles', getPercentiles_indexName - The name of the index to search.

$sel:queryString:GetPercentiles', getPercentiles_queryString - The search query string.

Request Lenses

getPercentiles_percents :: Lens' GetPercentiles (Maybe [Double])

The percentile groups returned.

getPercentiles_queryVersion :: Lens' GetPercentiles (Maybe Text)

The query version.

getPercentiles_aggregationField :: Lens' GetPercentiles (Maybe Text)

The field to aggregate.

getPercentiles_indexName :: Lens' GetPercentiles (Maybe Text)

The name of the index to search.

getPercentiles_queryString :: Lens' GetPercentiles Text

The search query string.

Destructuring the Response

data GetPercentilesResponse

See: newGetPercentilesResponse smart constructor.

Constructors

GetPercentilesResponse'

Fields

percentiles :: Maybe [PercentPair]

The percentile values of the aggregated fields.

httpStatus :: Int

The response's http status code.

newGetPercentilesResponse

Arguments

:: Int
$sel:httpStatus:GetPercentilesResponse'

-> GetPercentilesResponse

Create a value of GetPercentilesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:percentiles:GetPercentilesResponse', getPercentilesResponse_percentiles - The percentile values of the aggregated fields.

$sel:httpStatus:GetPercentilesResponse', getPercentilesResponse_httpStatus - The response's http status code.

Response Lenses

getPercentilesResponse_percentiles :: Lens' GetPercentilesResponse (Maybe [PercentPair])

The percentile values of the aggregated fields.

getPercentilesResponse_httpStatus :: Lens' GetPercentilesResponse Int

The response's http status code.