Network.AWS.IoT.GetTopicRuleDestination

Description

Gets information about a topic rule destination.

Requires permission to access the GetTopicRuleDestination action.

Creating a Request

data GetTopicRuleDestination

See: newGetTopicRuleDestination smart constructor.

Constructors

GetTopicRuleDestination'

Fields

arn :: Text

The ARN of the topic rule destination.

newGetTopicRuleDestination

Arguments

:: Text
$sel:arn:GetTopicRuleDestination'

-> GetTopicRuleDestination

Create a value of GetTopicRuleDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:GetTopicRuleDestination', getTopicRuleDestination_arn - The ARN of the topic rule destination.

Request Lenses

getTopicRuleDestination_arn :: Lens' GetTopicRuleDestination Text

The ARN of the topic rule destination.

Destructuring the Response

data GetTopicRuleDestinationResponse

See: newGetTopicRuleDestinationResponse smart constructor.

Constructors

GetTopicRuleDestinationResponse'

Fields

topicRuleDestination :: Maybe TopicRuleDestination

The topic rule destination.

httpStatus :: Int

The response's http status code.

newGetTopicRuleDestinationResponse

Arguments

:: Int
$sel:httpStatus:GetTopicRuleDestinationResponse'

-> GetTopicRuleDestinationResponse

Create a value of GetTopicRuleDestinationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:topicRuleDestination:GetTopicRuleDestinationResponse', getTopicRuleDestinationResponse_topicRuleDestination - The topic rule destination.

$sel:httpStatus:GetTopicRuleDestinationResponse', getTopicRuleDestinationResponse_httpStatus - The response's http status code.

Response Lenses

getTopicRuleDestinationResponse_topicRuleDestination :: Lens' GetTopicRuleDestinationResponse (Maybe TopicRuleDestination)

The topic rule destination.

getTopicRuleDestinationResponse_httpStatus :: Lens' GetTopicRuleDestinationResponse Int

The response's http status code.