Network.AWS.IoT.ListActiveViolations

Description

Lists the active violations for a given Device Defender security profile.

Requires permission to access the ListActiveViolations action.

This operation returns paginated results.

Creating a Request

data ListActiveViolations

See: newListActiveViolations smart constructor.

Constructors

ListActiveViolations'

Fields

nextToken :: Maybe Text

The token for the next set of results.

listSuppressedAlerts :: Maybe Bool

A list of all suppressed alerts.

behaviorCriteriaType :: Maybe BehaviorCriteriaType

The criteria for a behavior.

securityProfileName :: Maybe Text

The name of the Device Defender security profile for which violations are listed.

thingName :: Maybe Text

The name of the thing whose active violations are listed.

verificationState :: Maybe VerificationState

The verification state of the violation (detect alarm).

maxResults :: Maybe Natural

The maximum number of results to return at one time.

newListActiveViolations :: ListActiveViolations

Create a value of ListActiveViolations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListActiveViolations', listActiveViolations_nextToken - The token for the next set of results.

$sel:listSuppressedAlerts:ListActiveViolations', listActiveViolations_listSuppressedAlerts - A list of all suppressed alerts.

$sel:behaviorCriteriaType:ListActiveViolations', listActiveViolations_behaviorCriteriaType - The criteria for a behavior.

$sel:securityProfileName:ListActiveViolations', listActiveViolations_securityProfileName - The name of the Device Defender security profile for which violations are listed.

$sel:thingName:ListActiveViolations', listActiveViolations_thingName - The name of the thing whose active violations are listed.

$sel:verificationState:ListActiveViolations', listActiveViolations_verificationState - The verification state of the violation (detect alarm).

$sel:maxResults:ListActiveViolations', listActiveViolations_maxResults - The maximum number of results to return at one time.

Request Lenses

listActiveViolations_nextToken :: Lens' ListActiveViolations (Maybe Text)

The token for the next set of results.

listActiveViolations_listSuppressedAlerts :: Lens' ListActiveViolations (Maybe Bool)

A list of all suppressed alerts.

listActiveViolations_behaviorCriteriaType :: Lens' ListActiveViolations (Maybe BehaviorCriteriaType)

The criteria for a behavior.

listActiveViolations_securityProfileName :: Lens' ListActiveViolations (Maybe Text)

The name of the Device Defender security profile for which violations are listed.

listActiveViolations_thingName :: Lens' ListActiveViolations (Maybe Text)

The name of the thing whose active violations are listed.

listActiveViolations_verificationState :: Lens' ListActiveViolations (Maybe VerificationState)

The verification state of the violation (detect alarm).

listActiveViolations_maxResults :: Lens' ListActiveViolations (Maybe Natural)

The maximum number of results to return at one time.

Destructuring the Response

data ListActiveViolationsResponse

See: newListActiveViolationsResponse smart constructor.

Constructors

ListActiveViolationsResponse'

Fields

activeViolations :: Maybe [ActiveViolation]

The list of active violations.

nextToken :: Maybe Text

A token that can be used to retrieve the next set of results, or null if there are no additional results.

httpStatus :: Int

The response's http status code.

newListActiveViolationsResponse

Arguments

:: Int
$sel:httpStatus:ListActiveViolationsResponse'

-> ListActiveViolationsResponse

Create a value of ListActiveViolationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:activeViolations:ListActiveViolationsResponse', listActiveViolationsResponse_activeViolations - The list of active violations.

$sel:nextToken:ListActiveViolations', listActiveViolationsResponse_nextToken - A token that can be used to retrieve the next set of results, or null if there are no additional results.

$sel:httpStatus:ListActiveViolationsResponse', listActiveViolationsResponse_httpStatus - The response's http status code.

Response Lenses

listActiveViolationsResponse_activeViolations :: Lens' ListActiveViolationsResponse (Maybe [ActiveViolation])

The list of active violations.

listActiveViolationsResponse_nextToken :: Lens' ListActiveViolationsResponse (Maybe Text)

A token that can be used to retrieve the next set of results, or null if there are no additional results.

listActiveViolationsResponse_httpStatus :: Lens' ListActiveViolationsResponse Int

The response's http status code.