Network.AWS.IoT.ListAuditSuppressions

Description

Lists your Device Defender audit listings.

Requires permission to access the ListAuditSuppressions action.

This operation returns paginated results.

Creating a Request

data ListAuditSuppressions

See: newListAuditSuppressions smart constructor.

Constructors

ListAuditSuppressions'

Fields

nextToken :: Maybe Text

The token for the next set of results.

maxResults :: Maybe Natural

The maximum number of results to return at one time. The default is 25.

resourceIdentifier :: Maybe ResourceIdentifier

checkName :: Maybe Text

ascendingOrder :: Maybe Bool

Determines whether suppressions are listed in ascending order by expiration date or not. If parameter isn't provided, ascendingOrder=true.

newListAuditSuppressions :: ListAuditSuppressions

Create a value of ListAuditSuppressions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAuditSuppressions', listAuditSuppressions_nextToken - The token for the next set of results.

$sel:maxResults:ListAuditSuppressions', listAuditSuppressions_maxResults - The maximum number of results to return at one time. The default is 25.

$sel:resourceIdentifier:ListAuditSuppressions', listAuditSuppressions_resourceIdentifier - Undocumented member.

$sel:checkName:ListAuditSuppressions', listAuditSuppressions_checkName - Undocumented member.

$sel:ascendingOrder:ListAuditSuppressions', listAuditSuppressions_ascendingOrder - Determines whether suppressions are listed in ascending order by expiration date or not. If parameter isn't provided, ascendingOrder=true.

Request Lenses

listAuditSuppressions_nextToken :: Lens' ListAuditSuppressions (Maybe Text)

The token for the next set of results.

listAuditSuppressions_maxResults :: Lens' ListAuditSuppressions (Maybe Natural)

The maximum number of results to return at one time. The default is 25.

listAuditSuppressions_resourceIdentifier :: Lens' ListAuditSuppressions (Maybe ResourceIdentifier)

Undocumented member.

listAuditSuppressions_checkName :: Lens' ListAuditSuppressions (Maybe Text)

Undocumented member.

listAuditSuppressions_ascendingOrder :: Lens' ListAuditSuppressions (Maybe Bool)

Determines whether suppressions are listed in ascending order by expiration date or not. If parameter isn't provided, ascendingOrder=true.

Destructuring the Response

data ListAuditSuppressionsResponse

See: newListAuditSuppressionsResponse smart constructor.

Constructors

ListAuditSuppressionsResponse'

Fields

nextToken :: Maybe Text

A token that can be used to retrieve the next set of results, or null if there are no additional results.

suppressions :: Maybe [AuditSuppression]

List of audit suppressions.

httpStatus :: Int

The response's http status code.

newListAuditSuppressionsResponse

Arguments

:: Int
$sel:httpStatus:ListAuditSuppressionsResponse'

-> ListAuditSuppressionsResponse

Create a value of ListAuditSuppressionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAuditSuppressions', listAuditSuppressionsResponse_nextToken - A token that can be used to retrieve the next set of results, or null if there are no additional results.

$sel:suppressions:ListAuditSuppressionsResponse', listAuditSuppressionsResponse_suppressions - List of audit suppressions.

$sel:httpStatus:ListAuditSuppressionsResponse', listAuditSuppressionsResponse_httpStatus - The response's http status code.

Response Lenses

listAuditSuppressionsResponse_nextToken :: Lens' ListAuditSuppressionsResponse (Maybe Text)

A token that can be used to retrieve the next set of results, or null if there are no additional results.

listAuditSuppressionsResponse_suppressions :: Lens' ListAuditSuppressionsResponse (Maybe [AuditSuppression])

List of audit suppressions.

listAuditSuppressionsResponse_httpStatus :: Lens' ListAuditSuppressionsResponse Int

The response's http status code.