Network.AWS.IoT.ListAuditTasks

Description

Lists the Device Defender audits that have been performed during a given time period.

Requires permission to access the ListAuditTasks action.

This operation returns paginated results.

Creating a Request

data ListAuditTasks

See: newListAuditTasks smart constructor.

Constructors

ListAuditTasks'

Fields

nextToken :: Maybe Text

The token for the next set of results.

maxResults :: Maybe Natural

The maximum number of results to return at one time. The default is 25.

taskStatus :: Maybe AuditTaskStatus

A filter to limit the output to audits with the specified completion status: can be one of "IN_PROGRESS", "COMPLETED", "FAILED", or "CANCELED".

taskType :: Maybe AuditTaskType

A filter to limit the output to the specified type of audit: can be one of "ON_DEMAND_AUDIT_TASK" or "SCHEDULED__AUDIT_TASK".

startTime :: POSIX

The beginning of the time period. Audit information is retained for a limited time (90 days). Requesting a start time prior to what is retained results in an "InvalidRequestException".

endTime :: POSIX

The end of the time period.

newListAuditTasks

Arguments

:: UTCTime
$sel:startTime:ListAuditTasks'

-> UTCTime
$sel:endTime:ListAuditTasks'

-> ListAuditTasks

Create a value of ListAuditTasks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAuditTasks', listAuditTasks_nextToken - The token for the next set of results.

$sel:maxResults:ListAuditTasks', listAuditTasks_maxResults - The maximum number of results to return at one time. The default is 25.

$sel:taskStatus:ListAuditTasks', listAuditTasks_taskStatus - A filter to limit the output to audits with the specified completion status: can be one of "IN_PROGRESS", "COMPLETED", "FAILED", or "CANCELED".

$sel:taskType:ListAuditTasks', listAuditTasks_taskType - A filter to limit the output to the specified type of audit: can be one of "ON_DEMAND_AUDIT_TASK" or "SCHEDULED__AUDIT_TASK".

$sel:startTime:ListAuditTasks', listAuditTasks_startTime - The beginning of the time period. Audit information is retained for a limited time (90 days). Requesting a start time prior to what is retained results in an "InvalidRequestException".

$sel:endTime:ListAuditTasks', listAuditTasks_endTime - The end of the time period.

Request Lenses

listAuditTasks_nextToken :: Lens' ListAuditTasks (Maybe Text)

The token for the next set of results.

listAuditTasks_maxResults :: Lens' ListAuditTasks (Maybe Natural)

The maximum number of results to return at one time. The default is 25.

listAuditTasks_taskStatus :: Lens' ListAuditTasks (Maybe AuditTaskStatus)

A filter to limit the output to audits with the specified completion status: can be one of "IN_PROGRESS", "COMPLETED", "FAILED", or "CANCELED".

listAuditTasks_taskType :: Lens' ListAuditTasks (Maybe AuditTaskType)

A filter to limit the output to the specified type of audit: can be one of "ON_DEMAND_AUDIT_TASK" or "SCHEDULED__AUDIT_TASK".

listAuditTasks_startTime :: Lens' ListAuditTasks UTCTime

The beginning of the time period. Audit information is retained for a limited time (90 days). Requesting a start time prior to what is retained results in an "InvalidRequestException".

listAuditTasks_endTime :: Lens' ListAuditTasks UTCTime

The end of the time period.

Destructuring the Response

data ListAuditTasksResponse

See: newListAuditTasksResponse smart constructor.

Constructors

ListAuditTasksResponse'

Fields

nextToken :: Maybe Text

A token that can be used to retrieve the next set of results, or null if there are no additional results.

tasks :: Maybe [AuditTaskMetadata]

The audits that were performed during the specified time period.

httpStatus :: Int

The response's http status code.

newListAuditTasksResponse

Arguments

:: Int
$sel:httpStatus:ListAuditTasksResponse'

-> ListAuditTasksResponse

Create a value of ListAuditTasksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAuditTasks', listAuditTasksResponse_nextToken - A token that can be used to retrieve the next set of results, or null if there are no additional results.

$sel:tasks:ListAuditTasksResponse', listAuditTasksResponse_tasks - The audits that were performed during the specified time period.

$sel:httpStatus:ListAuditTasksResponse', listAuditTasksResponse_httpStatus - The response's http status code.

Response Lenses

listAuditTasksResponse_nextToken :: Lens' ListAuditTasksResponse (Maybe Text)

A token that can be used to retrieve the next set of results, or null if there are no additional results.

listAuditTasksResponse_tasks :: Lens' ListAuditTasksResponse (Maybe [AuditTaskMetadata])

The audits that were performed during the specified time period.

listAuditTasksResponse_httpStatus :: Lens' ListAuditTasksResponse Int

The response's http status code.