Network.AWS.IoT.ListCertificates

Description

Lists the certificates registered in your Amazon Web Services account.

The results are paginated with a default page size of 25. You can use the returned marker to retrieve additional results.

Requires permission to access the ListCertificates action.

This operation returns paginated results.

Creating a Request

data ListCertificates

The input for the ListCertificates operation.

See: newListCertificates smart constructor.

Constructors

ListCertificates'

Fields

pageSize :: Maybe Natural

The result page size.

ascendingOrder :: Maybe Bool

Specifies the order for results. If True, the results are returned in ascending order, based on the creation date.

marker :: Maybe Text

The marker for the next set of results.

newListCertificates :: ListCertificates

Create a value of ListCertificates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pageSize:ListCertificates', listCertificates_pageSize - The result page size.

$sel:ascendingOrder:ListCertificates', listCertificates_ascendingOrder - Specifies the order for results. If True, the results are returned in ascending order, based on the creation date.

$sel:marker:ListCertificates', listCertificates_marker - The marker for the next set of results.

Request Lenses

listCertificates_pageSize :: Lens' ListCertificates (Maybe Natural)

The result page size.

listCertificates_ascendingOrder :: Lens' ListCertificates (Maybe Bool)

Specifies the order for results. If True, the results are returned in ascending order, based on the creation date.

listCertificates_marker :: Lens' ListCertificates (Maybe Text)

The marker for the next set of results.

Destructuring the Response

data ListCertificatesResponse

The output of the ListCertificates operation.

See: newListCertificatesResponse smart constructor.

Constructors

ListCertificatesResponse'

Fields

nextMarker :: Maybe Text

The marker for the next set of results, or null if there are no additional results.

certificates :: Maybe [Certificate]

The descriptions of the certificates.

httpStatus :: Int

The response's http status code.

newListCertificatesResponse

Arguments

:: Int
$sel:httpStatus:ListCertificatesResponse'

-> ListCertificatesResponse

Create a value of ListCertificatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextMarker:ListCertificatesResponse', listCertificatesResponse_nextMarker - The marker for the next set of results, or null if there are no additional results.

$sel:certificates:ListCertificatesResponse', listCertificatesResponse_certificates - The descriptions of the certificates.

$sel:httpStatus:ListCertificatesResponse', listCertificatesResponse_httpStatus - The response's http status code.

Response Lenses

listCertificatesResponse_nextMarker :: Lens' ListCertificatesResponse (Maybe Text)

The marker for the next set of results, or null if there are no additional results.

listCertificatesResponse_certificates :: Lens' ListCertificatesResponse (Maybe [Certificate])

The descriptions of the certificates.

listCertificatesResponse_httpStatus :: Lens' ListCertificatesResponse Int

The response's http status code.