Network.AWS.IoT.ListDetectMitigationActionsExecutions

Description

Lists mitigation actions executions for a Device Defender ML Detect Security Profile.

Requires permission to access the ListDetectMitigationActionsExecutions action.

This operation returns paginated results.

Creating a Request

data ListDetectMitigationActionsExecutions

See: newListDetectMitigationActionsExecutions smart constructor.

Constructors

ListDetectMitigationActionsExecutions'

Fields

startTime :: Maybe POSIX

A filter to limit results to those found after the specified time. You must specify either the startTime and endTime or the taskId, but not both.

taskId :: Maybe Text

The unique identifier of the task.

violationId :: Maybe Text

The unique identifier of the violation.

nextToken :: Maybe Text

The token for the next set of results.

endTime :: Maybe POSIX

The end of the time period for which ML Detect mitigation actions executions are returned.

thingName :: Maybe Text

The name of the thing whose mitigation actions are listed.

maxResults :: Maybe Natural

The maximum number of results to return at one time. The default is 25.

newListDetectMitigationActionsExecutions :: ListDetectMitigationActionsExecutions

Create a value of ListDetectMitigationActionsExecutions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startTime:ListDetectMitigationActionsExecutions', listDetectMitigationActionsExecutions_startTime - A filter to limit results to those found after the specified time. You must specify either the startTime and endTime or the taskId, but not both.

$sel:taskId:ListDetectMitigationActionsExecutions', listDetectMitigationActionsExecutions_taskId - The unique identifier of the task.

$sel:violationId:ListDetectMitigationActionsExecutions', listDetectMitigationActionsExecutions_violationId - The unique identifier of the violation.

$sel:nextToken:ListDetectMitigationActionsExecutions', listDetectMitigationActionsExecutions_nextToken - The token for the next set of results.

$sel:endTime:ListDetectMitigationActionsExecutions', listDetectMitigationActionsExecutions_endTime - The end of the time period for which ML Detect mitigation actions executions are returned.

$sel:thingName:ListDetectMitigationActionsExecutions', listDetectMitigationActionsExecutions_thingName - The name of the thing whose mitigation actions are listed.

$sel:maxResults:ListDetectMitigationActionsExecutions', listDetectMitigationActionsExecutions_maxResults - The maximum number of results to return at one time. The default is 25.

Request Lenses

listDetectMitigationActionsExecutions_startTime :: Lens' ListDetectMitigationActionsExecutions (Maybe UTCTime)

A filter to limit results to those found after the specified time. You must specify either the startTime and endTime or the taskId, but not both.

listDetectMitigationActionsExecutions_taskId :: Lens' ListDetectMitigationActionsExecutions (Maybe Text)

The unique identifier of the task.

listDetectMitigationActionsExecutions_violationId :: Lens' ListDetectMitigationActionsExecutions (Maybe Text)

The unique identifier of the violation.

listDetectMitigationActionsExecutions_nextToken :: Lens' ListDetectMitigationActionsExecutions (Maybe Text)

The token for the next set of results.

listDetectMitigationActionsExecutions_endTime :: Lens' ListDetectMitigationActionsExecutions (Maybe UTCTime)

The end of the time period for which ML Detect mitigation actions executions are returned.

listDetectMitigationActionsExecutions_thingName :: Lens' ListDetectMitigationActionsExecutions (Maybe Text)

The name of the thing whose mitigation actions are listed.

listDetectMitigationActionsExecutions_maxResults :: Lens' ListDetectMitigationActionsExecutions (Maybe Natural)

The maximum number of results to return at one time. The default is 25.

Destructuring the Response

data ListDetectMitigationActionsExecutionsResponse

See: newListDetectMitigationActionsExecutionsResponse smart constructor.

Constructors

ListDetectMitigationActionsExecutionsResponse'

Fields

actionsExecutions :: Maybe [DetectMitigationActionExecution]

List of actions executions.

nextToken :: Maybe Text

A token that can be used to retrieve the next set of results, or null if there are no additional results.

httpStatus :: Int

The response's http status code.

newListDetectMitigationActionsExecutionsResponse

Arguments

:: Int
$sel:httpStatus:ListDetectMitigationActionsExecutionsResponse'

-> ListDetectMitigationActionsExecutionsResponse

Create a value of ListDetectMitigationActionsExecutionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actionsExecutions:ListDetectMitigationActionsExecutionsResponse', listDetectMitigationActionsExecutionsResponse_actionsExecutions - List of actions executions.

$sel:nextToken:ListDetectMitigationActionsExecutions', listDetectMitigationActionsExecutionsResponse_nextToken - A token that can be used to retrieve the next set of results, or null if there are no additional results.

$sel:httpStatus:ListDetectMitigationActionsExecutionsResponse', listDetectMitigationActionsExecutionsResponse_httpStatus - The response's http status code.

Response Lenses

listDetectMitigationActionsExecutionsResponse_actionsExecutions :: Lens' ListDetectMitigationActionsExecutionsResponse (Maybe [DetectMitigationActionExecution])

List of actions executions.

listDetectMitigationActionsExecutionsResponse_nextToken :: Lens' ListDetectMitigationActionsExecutionsResponse (Maybe Text)

A token that can be used to retrieve the next set of results, or null if there are no additional results.

listDetectMitigationActionsExecutionsResponse_httpStatus :: Lens' ListDetectMitigationActionsExecutionsResponse Int

The response's http status code.