Network.AWS.IoT.ListDetectMitigationActionsTasks

Description

List of Device Defender ML Detect mitigation actions tasks.

Requires permission to access the ListDetectMitigationActionsTasks action.

This operation returns paginated results.

Creating a Request

data ListDetectMitigationActionsTasks

See: newListDetectMitigationActionsTasks smart constructor.

Constructors

ListDetectMitigationActionsTasks'

Fields

nextToken :: Maybe Text

The token for the next set of results.

maxResults :: Maybe Natural

The maximum number of results to return at one time. The default is 25.

startTime :: POSIX

A filter to limit results to those found after the specified time. You must specify either the startTime and endTime or the taskId, but not both.

endTime :: POSIX

The end of the time period for which ML Detect mitigation actions tasks are returned.

newListDetectMitigationActionsTasks

Arguments

:: UTCTime
$sel:startTime:ListDetectMitigationActionsTasks'

-> UTCTime
$sel:endTime:ListDetectMitigationActionsTasks'

-> ListDetectMitigationActionsTasks

Create a value of ListDetectMitigationActionsTasks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDetectMitigationActionsTasks', listDetectMitigationActionsTasks_nextToken - The token for the next set of results.

$sel:maxResults:ListDetectMitigationActionsTasks', listDetectMitigationActionsTasks_maxResults - The maximum number of results to return at one time. The default is 25.

$sel:startTime:ListDetectMitigationActionsTasks', listDetectMitigationActionsTasks_startTime - A filter to limit results to those found after the specified time. You must specify either the startTime and endTime or the taskId, but not both.

$sel:endTime:ListDetectMitigationActionsTasks', listDetectMitigationActionsTasks_endTime - The end of the time period for which ML Detect mitigation actions tasks are returned.

Request Lenses

listDetectMitigationActionsTasks_nextToken :: Lens' ListDetectMitigationActionsTasks (Maybe Text)

The token for the next set of results.

listDetectMitigationActionsTasks_maxResults :: Lens' ListDetectMitigationActionsTasks (Maybe Natural)

The maximum number of results to return at one time. The default is 25.

listDetectMitigationActionsTasks_startTime :: Lens' ListDetectMitigationActionsTasks UTCTime

A filter to limit results to those found after the specified time. You must specify either the startTime and endTime or the taskId, but not both.

listDetectMitigationActionsTasks_endTime :: Lens' ListDetectMitigationActionsTasks UTCTime

The end of the time period for which ML Detect mitigation actions tasks are returned.

Destructuring the Response

data ListDetectMitigationActionsTasksResponse

See: newListDetectMitigationActionsTasksResponse smart constructor.

Constructors

ListDetectMitigationActionsTasksResponse'

Fields

tasks :: Maybe [DetectMitigationActionsTaskSummary]

The collection of ML Detect mitigation tasks that matched the filter criteria.

nextToken :: Maybe Text

A token that can be used to retrieve the next set of results, or null if there are no additional results.

httpStatus :: Int

The response's http status code.

newListDetectMitigationActionsTasksResponse

Arguments

:: Int
$sel:httpStatus:ListDetectMitigationActionsTasksResponse'

-> ListDetectMitigationActionsTasksResponse

Create a value of ListDetectMitigationActionsTasksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tasks:ListDetectMitigationActionsTasksResponse', listDetectMitigationActionsTasksResponse_tasks - The collection of ML Detect mitigation tasks that matched the filter criteria.

$sel:nextToken:ListDetectMitigationActionsTasks', listDetectMitigationActionsTasksResponse_nextToken - A token that can be used to retrieve the next set of results, or null if there are no additional results.

$sel:httpStatus:ListDetectMitigationActionsTasksResponse', listDetectMitigationActionsTasksResponse_httpStatus - The response's http status code.

Response Lenses

listDetectMitigationActionsTasksResponse_tasks :: Lens' ListDetectMitigationActionsTasksResponse (Maybe [DetectMitigationActionsTaskSummary])

The collection of ML Detect mitigation tasks that matched the filter criteria.

listDetectMitigationActionsTasksResponse_nextToken :: Lens' ListDetectMitigationActionsTasksResponse (Maybe Text)

A token that can be used to retrieve the next set of results, or null if there are no additional results.

listDetectMitigationActionsTasksResponse_httpStatus :: Lens' ListDetectMitigationActionsTasksResponse Int

The response's http status code.