Network.AWS.IoT.ListJobExecutionsForThing

Description

Lists the job executions for the specified thing.

Requires permission to access the ListJobExecutionsForThing action.

This operation returns paginated results.

Creating a Request

data ListJobExecutionsForThing

See: newListJobExecutionsForThing smart constructor.

Constructors

ListJobExecutionsForThing'

Fields

status :: Maybe JobExecutionStatus

An optional filter that lets you search for jobs that have the specified status.

namespaceId :: Maybe Text

The namespace used to indicate that a job is a customer-managed job.

When you specify a value for this parameter, Amazon Web Services IoT Core sends jobs notifications to MQTT topics that contain the value in the following format.

$aws/things/THING_NAME/jobs/JOB_ID/notify-namespace-NAMESPACE_ID/

The namespaceId feature is in public preview.

nextToken :: Maybe Text

The token to retrieve the next set of results.

maxResults :: Maybe Natural

The maximum number of results to be returned per request.

thingName :: Text

The thing name.

newListJobExecutionsForThing

Arguments

:: Text
$sel:thingName:ListJobExecutionsForThing'

-> ListJobExecutionsForThing

Create a value of ListJobExecutionsForThing with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListJobExecutionsForThing', listJobExecutionsForThing_status - An optional filter that lets you search for jobs that have the specified status.

$sel:namespaceId:ListJobExecutionsForThing', listJobExecutionsForThing_namespaceId - The namespace used to indicate that a job is a customer-managed job.

When you specify a value for this parameter, Amazon Web Services IoT Core sends jobs notifications to MQTT topics that contain the value in the following format.

$aws/things/THING_NAME/jobs/JOB_ID/notify-namespace-NAMESPACE_ID/

The namespaceId feature is in public preview.

$sel:nextToken:ListJobExecutionsForThing', listJobExecutionsForThing_nextToken - The token to retrieve the next set of results.

$sel:maxResults:ListJobExecutionsForThing', listJobExecutionsForThing_maxResults - The maximum number of results to be returned per request.

$sel:thingName:ListJobExecutionsForThing', listJobExecutionsForThing_thingName - The thing name.

Request Lenses

listJobExecutionsForThing_status :: Lens' ListJobExecutionsForThing (Maybe JobExecutionStatus)

An optional filter that lets you search for jobs that have the specified status.

listJobExecutionsForThing_namespaceId :: Lens' ListJobExecutionsForThing (Maybe Text)

The namespace used to indicate that a job is a customer-managed job.

When you specify a value for this parameter, Amazon Web Services IoT Core sends jobs notifications to MQTT topics that contain the value in the following format.

$aws/things/THING_NAME/jobs/JOB_ID/notify-namespace-NAMESPACE_ID/

The namespaceId feature is in public preview.

listJobExecutionsForThing_nextToken :: Lens' ListJobExecutionsForThing (Maybe Text)

The token to retrieve the next set of results.

listJobExecutionsForThing_maxResults :: Lens' ListJobExecutionsForThing (Maybe Natural)

The maximum number of results to be returned per request.

listJobExecutionsForThing_thingName :: Lens' ListJobExecutionsForThing Text

The thing name.

Destructuring the Response

data ListJobExecutionsForThingResponse

See: newListJobExecutionsForThingResponse smart constructor.

Constructors

ListJobExecutionsForThingResponse'

Fields

executionSummaries :: Maybe [JobExecutionSummaryForThing]

A list of job execution summaries.

nextToken :: Maybe Text

The token for the next set of results, or null if there are no additional results.

httpStatus :: Int

The response's http status code.

newListJobExecutionsForThingResponse

Arguments

:: Int
$sel:httpStatus:ListJobExecutionsForThingResponse'

-> ListJobExecutionsForThingResponse

Create a value of ListJobExecutionsForThingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:executionSummaries:ListJobExecutionsForThingResponse', listJobExecutionsForThingResponse_executionSummaries - A list of job execution summaries.

$sel:nextToken:ListJobExecutionsForThing', listJobExecutionsForThingResponse_nextToken - The token for the next set of results, or null if there are no additional results.

$sel:httpStatus:ListJobExecutionsForThingResponse', listJobExecutionsForThingResponse_httpStatus - The response's http status code.

Response Lenses

listJobExecutionsForThingResponse_executionSummaries :: Lens' ListJobExecutionsForThingResponse (Maybe [JobExecutionSummaryForThing])

A list of job execution summaries.

listJobExecutionsForThingResponse_nextToken :: Lens' ListJobExecutionsForThingResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

listJobExecutionsForThingResponse_httpStatus :: Lens' ListJobExecutionsForThingResponse Int

The response's http status code.