Network.AWS.IoT.ListPrincipalThings

Description

Lists the things associated with the specified principal. A principal can be X.509 certificates, IAM users, groups, and roles, Amazon Cognito identities or federated identities.

Requires permission to access the ListPrincipalThings action.

This operation returns paginated results.

Creating a Request

data ListPrincipalThings

The input for the ListPrincipalThings operation.

See: newListPrincipalThings smart constructor.

Constructors

ListPrincipalThings'

Fields

nextToken :: Maybe Text

To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

maxResults :: Maybe Natural

The maximum number of results to return in this operation.

principal :: Text

The principal.

newListPrincipalThings

Arguments

:: Text
$sel:principal:ListPrincipalThings'

-> ListPrincipalThings

Create a value of ListPrincipalThings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPrincipalThings', listPrincipalThings_nextToken - To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

$sel:maxResults:ListPrincipalThings', listPrincipalThings_maxResults - The maximum number of results to return in this operation.

$sel:principal:ListPrincipalThings', listPrincipalThings_principal - The principal.

Request Lenses

listPrincipalThings_nextToken :: Lens' ListPrincipalThings (Maybe Text)

To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

listPrincipalThings_maxResults :: Lens' ListPrincipalThings (Maybe Natural)

The maximum number of results to return in this operation.

listPrincipalThings_principal :: Lens' ListPrincipalThings Text

The principal.

Destructuring the Response

data ListPrincipalThingsResponse

The output from the ListPrincipalThings operation.

See: newListPrincipalThingsResponse smart constructor.

Constructors

ListPrincipalThingsResponse'

Fields

nextToken :: Maybe Text

The token to use to get the next set of results, or null if there are no additional results.

things :: Maybe [Text]

The things.

httpStatus :: Int

The response's http status code.

newListPrincipalThingsResponse

Arguments

:: Int
$sel:httpStatus:ListPrincipalThingsResponse'

-> ListPrincipalThingsResponse

Create a value of ListPrincipalThingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPrincipalThings', listPrincipalThingsResponse_nextToken - The token to use to get the next set of results, or null if there are no additional results.

$sel:things:ListPrincipalThingsResponse', listPrincipalThingsResponse_things - The things.

$sel:httpStatus:ListPrincipalThingsResponse', listPrincipalThingsResponse_httpStatus - The response's http status code.

Response Lenses

listPrincipalThingsResponse_nextToken :: Lens' ListPrincipalThingsResponse (Maybe Text)

The token to use to get the next set of results, or null if there are no additional results.

listPrincipalThingsResponse_things :: Lens' ListPrincipalThingsResponse (Maybe [Text])

The things.

listPrincipalThingsResponse_httpStatus :: Lens' ListPrincipalThingsResponse Int

The response's http status code.