Network.AWS.IoT.ListThingGroupsForThing

Description

List the thing groups to which the specified thing belongs.

Requires permission to access the ListThingGroupsForThing action.

This operation returns paginated results.

Creating a Request

data ListThingGroupsForThing

See: newListThingGroupsForThing smart constructor.

Constructors

ListThingGroupsForThing'

Fields

nextToken :: Maybe Text

To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

maxResults :: Maybe Natural

The maximum number of results to return at one time.

thingName :: Text

The thing name.

newListThingGroupsForThing

Arguments

:: Text
$sel:thingName:ListThingGroupsForThing'

-> ListThingGroupsForThing

Create a value of ListThingGroupsForThing with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListThingGroupsForThing', listThingGroupsForThing_nextToken - To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

$sel:maxResults:ListThingGroupsForThing', listThingGroupsForThing_maxResults - The maximum number of results to return at one time.

$sel:thingName:ListThingGroupsForThing', listThingGroupsForThing_thingName - The thing name.

Request Lenses

listThingGroupsForThing_nextToken :: Lens' ListThingGroupsForThing (Maybe Text)

To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

listThingGroupsForThing_maxResults :: Lens' ListThingGroupsForThing (Maybe Natural)

The maximum number of results to return at one time.

listThingGroupsForThing_thingName :: Lens' ListThingGroupsForThing Text

The thing name.

Destructuring the Response

data ListThingGroupsForThingResponse

See: newListThingGroupsForThingResponse smart constructor.

Constructors

ListThingGroupsForThingResponse'

Fields

nextToken :: Maybe Text

The token to use to get the next set of results, or null if there are no additional results.

thingGroups :: Maybe [GroupNameAndArn]

The thing groups.

httpStatus :: Int

The response's http status code.

newListThingGroupsForThingResponse

Arguments

:: Int
$sel:httpStatus:ListThingGroupsForThingResponse'

-> ListThingGroupsForThingResponse

Create a value of ListThingGroupsForThingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListThingGroupsForThing', listThingGroupsForThingResponse_nextToken - The token to use to get the next set of results, or null if there are no additional results.

$sel:thingGroups:ListThingGroupsForThingResponse', listThingGroupsForThingResponse_thingGroups - The thing groups.

$sel:httpStatus:ListThingGroupsForThingResponse', listThingGroupsForThingResponse_httpStatus - The response's http status code.

Response Lenses

listThingGroupsForThingResponse_nextToken :: Lens' ListThingGroupsForThingResponse (Maybe Text)

The token to use to get the next set of results, or null if there are no additional results.

listThingGroupsForThingResponse_thingGroups :: Lens' ListThingGroupsForThingResponse (Maybe [GroupNameAndArn])

The thing groups.

listThingGroupsForThingResponse_httpStatus :: Lens' ListThingGroupsForThingResponse Int

The response's http status code.