Network.AWS.IoT.ListThingRegistrationTaskReports

Description

Information about the thing registration tasks.

This operation returns paginated results.

Creating a Request

data ListThingRegistrationTaskReports

See: newListThingRegistrationTaskReports smart constructor.

Constructors

ListThingRegistrationTaskReports'

Fields

nextToken :: Maybe Text

To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

maxResults :: Maybe Natural

The maximum number of results to return per request.

taskId :: Text

The id of the task.

reportType :: ReportType

The type of task report.

newListThingRegistrationTaskReports

Arguments

:: Text
$sel:taskId:ListThingRegistrationTaskReports'

-> ReportType
$sel:reportType:ListThingRegistrationTaskReports'

-> ListThingRegistrationTaskReports

Create a value of ListThingRegistrationTaskReports with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListThingRegistrationTaskReports', listThingRegistrationTaskReports_nextToken - To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

$sel:maxResults:ListThingRegistrationTaskReports', listThingRegistrationTaskReports_maxResults - The maximum number of results to return per request.

$sel:taskId:ListThingRegistrationTaskReports', listThingRegistrationTaskReports_taskId - The id of the task.

$sel:reportType:ListThingRegistrationTaskReports', listThingRegistrationTaskReports_reportType - The type of task report.

Request Lenses

listThingRegistrationTaskReports_nextToken :: Lens' ListThingRegistrationTaskReports (Maybe Text)

To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

listThingRegistrationTaskReports_maxResults :: Lens' ListThingRegistrationTaskReports (Maybe Natural)

The maximum number of results to return per request.

listThingRegistrationTaskReports_taskId :: Lens' ListThingRegistrationTaskReports Text

The id of the task.

listThingRegistrationTaskReports_reportType :: Lens' ListThingRegistrationTaskReports ReportType

The type of task report.

Destructuring the Response

data ListThingRegistrationTaskReportsResponse

See: newListThingRegistrationTaskReportsResponse smart constructor.

Constructors

ListThingRegistrationTaskReportsResponse'

Fields

resourceLinks :: Maybe [Text]

Links to the task resources.

nextToken :: Maybe Text

The token to use to get the next set of results, or null if there are no additional results.

reportType :: Maybe ReportType

The type of task report.

httpStatus :: Int

The response's http status code.

newListThingRegistrationTaskReportsResponse

Arguments

:: Int
$sel:httpStatus:ListThingRegistrationTaskReportsResponse'

-> ListThingRegistrationTaskReportsResponse

Create a value of ListThingRegistrationTaskReportsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceLinks:ListThingRegistrationTaskReportsResponse', listThingRegistrationTaskReportsResponse_resourceLinks - Links to the task resources.

$sel:nextToken:ListThingRegistrationTaskReports', listThingRegistrationTaskReportsResponse_nextToken - The token to use to get the next set of results, or null if there are no additional results.

$sel:reportType:ListThingRegistrationTaskReports', listThingRegistrationTaskReportsResponse_reportType - The type of task report.

$sel:httpStatus:ListThingRegistrationTaskReportsResponse', listThingRegistrationTaskReportsResponse_httpStatus - The response's http status code.

Response Lenses

listThingRegistrationTaskReportsResponse_resourceLinks :: Lens' ListThingRegistrationTaskReportsResponse (Maybe [Text])

Links to the task resources.

listThingRegistrationTaskReportsResponse_nextToken :: Lens' ListThingRegistrationTaskReportsResponse (Maybe Text)

The token to use to get the next set of results, or null if there are no additional results.

listThingRegistrationTaskReportsResponse_reportType :: Lens' ListThingRegistrationTaskReportsResponse (Maybe ReportType)

The type of task report.

listThingRegistrationTaskReportsResponse_httpStatus :: Lens' ListThingRegistrationTaskReportsResponse Int

The response's http status code.