Network.AWS.IoT.ListThingTypes

Description

Lists the existing thing types.

Requires permission to access the ListThingTypes action.

This operation returns paginated results.

Creating a Request

data ListThingTypes

The input for the ListThingTypes operation.

See: newListThingTypes smart constructor.

Constructors

ListThingTypes'

Fields

thingTypeName :: Maybe Text

The name of the thing type.

nextToken :: Maybe Text

To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

maxResults :: Maybe Natural

The maximum number of results to return in this operation.

newListThingTypes :: ListThingTypes

Create a value of ListThingTypes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thingTypeName:ListThingTypes', listThingTypes_thingTypeName - The name of the thing type.

$sel:nextToken:ListThingTypes', listThingTypes_nextToken - To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

$sel:maxResults:ListThingTypes', listThingTypes_maxResults - The maximum number of results to return in this operation.

Request Lenses

listThingTypes_thingTypeName :: Lens' ListThingTypes (Maybe Text)

The name of the thing type.

listThingTypes_nextToken :: Lens' ListThingTypes (Maybe Text)

To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

listThingTypes_maxResults :: Lens' ListThingTypes (Maybe Natural)

The maximum number of results to return in this operation.

Destructuring the Response

data ListThingTypesResponse

The output for the ListThingTypes operation.

See: newListThingTypesResponse smart constructor.

Constructors

ListThingTypesResponse'

Fields

thingTypes :: Maybe [ThingTypeDefinition]

The thing types.

nextToken :: Maybe Text

The token for the next set of results. Will not be returned if operation has returned all results.

httpStatus :: Int

The response's http status code.

newListThingTypesResponse

Arguments

:: Int
$sel:httpStatus:ListThingTypesResponse'

-> ListThingTypesResponse

Create a value of ListThingTypesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thingTypes:ListThingTypesResponse', listThingTypesResponse_thingTypes - The thing types.

$sel:nextToken:ListThingTypes', listThingTypesResponse_nextToken - The token for the next set of results. Will not be returned if operation has returned all results.

$sel:httpStatus:ListThingTypesResponse', listThingTypesResponse_httpStatus - The response's http status code.

Response Lenses

listThingTypesResponse_thingTypes :: Lens' ListThingTypesResponse (Maybe [ThingTypeDefinition])

The thing types.

listThingTypesResponse_nextToken :: Lens' ListThingTypesResponse (Maybe Text)

The token for the next set of results. Will not be returned if operation has returned all results.

listThingTypesResponse_httpStatus :: Lens' ListThingTypesResponse Int

The response's http status code.