Network.AWS.IoT.ListThings

Description

Lists your things. Use the attributeName and attributeValue parameters to filter your things. For example, calling ListThings with attributeName=Color and attributeValue=Red retrieves all things in the registry that contain an attribute Color with the value Red.

Requires permission to access the ListThings action.

You will not be charged for calling this API if an Access denied error is returned. You will also not be charged if no attributes or pagination token was provided in request and no pagination token and no results were returned.

This operation returns paginated results.

Creating a Request

data ListThings

The input for the ListThings operation.

See: newListThings smart constructor.

Constructors

ListThings'

Fields

attributeValue :: Maybe Text

The attribute value used to search for things.

nextToken :: Maybe Text

To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

maxResults :: Maybe Natural

The maximum number of results to return in this operation.

attributeName :: Maybe Text

The attribute name used to search for things.

usePrefixAttributeValue :: Maybe Bool

When true, the action returns the thing resources with attribute values that start with the attributeValue provided.

When false, or not present, the action returns only the thing resources with attribute values that match the entire attributeValue provided.

thingTypeName :: Maybe Text

The name of the thing type used to search for things.

newListThings :: ListThings

Create a value of ListThings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeValue:ListThings', listThings_attributeValue - The attribute value used to search for things.

$sel:nextToken:ListThings', listThings_nextToken - To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

$sel:maxResults:ListThings', listThings_maxResults - The maximum number of results to return in this operation.

$sel:attributeName:ListThings', listThings_attributeName - The attribute name used to search for things.

$sel:usePrefixAttributeValue:ListThings', listThings_usePrefixAttributeValue - When true, the action returns the thing resources with attribute values that start with the attributeValue provided.

When false, or not present, the action returns only the thing resources with attribute values that match the entire attributeValue provided.

$sel:thingTypeName:ListThings', listThings_thingTypeName - The name of the thing type used to search for things.

Request Lenses

listThings_attributeValue :: Lens' ListThings (Maybe Text)

The attribute value used to search for things.

listThings_nextToken :: Lens' ListThings (Maybe Text)

To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

listThings_maxResults :: Lens' ListThings (Maybe Natural)

The maximum number of results to return in this operation.

listThings_attributeName :: Lens' ListThings (Maybe Text)

The attribute name used to search for things.

listThings_usePrefixAttributeValue :: Lens' ListThings (Maybe Bool)

When true, the action returns the thing resources with attribute values that start with the attributeValue provided.

When false, or not present, the action returns only the thing resources with attribute values that match the entire attributeValue provided.

listThings_thingTypeName :: Lens' ListThings (Maybe Text)

The name of the thing type used to search for things.

Destructuring the Response

data ListThingsResponse

The output from the ListThings operation.

See: newListThingsResponse smart constructor.

Constructors

ListThingsResponse'

Fields

nextToken :: Maybe Text

The token to use to get the next set of results. Will not be returned if operation has returned all results.

things :: Maybe [ThingAttribute]

The things.

httpStatus :: Int

The response's http status code.

newListThingsResponse

Arguments

:: Int
$sel:httpStatus:ListThingsResponse'

-> ListThingsResponse

Create a value of ListThingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListThings', listThingsResponse_nextToken - The token to use to get the next set of results. Will not be returned if operation has returned all results.

$sel:things:ListThingsResponse', listThingsResponse_things - The things.

$sel:httpStatus:ListThingsResponse', listThingsResponse_httpStatus - The response's http status code.

Response Lenses

listThingsResponse_nextToken :: Lens' ListThingsResponse (Maybe Text)

The token to use to get the next set of results. Will not be returned if operation has returned all results.

listThingsResponse_things :: Lens' ListThingsResponse (Maybe [ThingAttribute])

The things.

listThingsResponse_httpStatus :: Lens' ListThingsResponse Int

The response's http status code.