Network.AWS.IoT.RegisterCACertificate

Description

Registers a CA certificate with IoT. This CA certificate can then be used to sign device certificates, which can be then registered with IoT. You can register up to 10 CA certificates per Amazon Web Services account that have the same subject field. This enables you to have up to 10 certificate authorities sign your device certificates. If you have more than one CA certificate registered, make sure you pass the CA certificate when you register your device certificates with the RegisterCertificate action.

Requires permission to access the RegisterCACertificate action.

Creating a Request

data RegisterCACertificate

The input to the RegisterCACertificate operation.

See: newRegisterCACertificate smart constructor.

Constructors

RegisterCACertificate'

Fields

allowAutoRegistration :: Maybe Bool

Allows this CA certificate to be used for auto registration of device certificates.

setAsActive :: Maybe Bool

A boolean value that specifies if the CA certificate is set to active.

tags :: Maybe [Tag]

Metadata which can be used to manage the CA certificate.

For URI Request parameters use format: ...key1=value1&key2=value2...

For the CLI command-line parameter use format: &&tags "key1=value1&key2=value2..."

For the cli-input-json file use format: "tags": "key1=value1&key2=value2..."

registrationConfig :: Maybe RegistrationConfig

Information about the registration configuration.

caCertificate :: Text

The CA certificate.

verificationCertificate :: Text

The private key verification certificate.

newRegisterCACertificate

Arguments

:: Text
$sel:caCertificate:RegisterCACertificate'

-> Text
$sel:verificationCertificate:RegisterCACertificate'

-> RegisterCACertificate

Create a value of RegisterCACertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:allowAutoRegistration:RegisterCACertificate', registerCACertificate_allowAutoRegistration - Allows this CA certificate to be used for auto registration of device certificates.

$sel:setAsActive:RegisterCACertificate', registerCACertificate_setAsActive - A boolean value that specifies if the CA certificate is set to active.

$sel:tags:RegisterCACertificate', registerCACertificate_tags - Metadata which can be used to manage the CA certificate.

For URI Request parameters use format: ...key1=value1&key2=value2...

For the CLI command-line parameter use format: &&tags "key1=value1&key2=value2..."

For the cli-input-json file use format: "tags": "key1=value1&key2=value2..."

$sel:registrationConfig:RegisterCACertificate', registerCACertificate_registrationConfig - Information about the registration configuration.

$sel:caCertificate:RegisterCACertificate', registerCACertificate_caCertificate - The CA certificate.

$sel:verificationCertificate:RegisterCACertificate', registerCACertificate_verificationCertificate - The private key verification certificate.

Request Lenses

registerCACertificate_allowAutoRegistration :: Lens' RegisterCACertificate (Maybe Bool)

Allows this CA certificate to be used for auto registration of device certificates.

registerCACertificate_setAsActive :: Lens' RegisterCACertificate (Maybe Bool)

A boolean value that specifies if the CA certificate is set to active.

registerCACertificate_tags :: Lens' RegisterCACertificate (Maybe [Tag])

Metadata which can be used to manage the CA certificate.

For URI Request parameters use format: ...key1=value1&key2=value2...

For the CLI command-line parameter use format: &&tags "key1=value1&key2=value2..."

For the cli-input-json file use format: "tags": "key1=value1&key2=value2..."

registerCACertificate_registrationConfig :: Lens' RegisterCACertificate (Maybe RegistrationConfig)

Information about the registration configuration.

registerCACertificate_caCertificate :: Lens' RegisterCACertificate Text

The CA certificate.

registerCACertificate_verificationCertificate :: Lens' RegisterCACertificate Text

The private key verification certificate.

Destructuring the Response

data RegisterCACertificateResponse

The output from the RegisterCACertificateResponse operation.

See: newRegisterCACertificateResponse smart constructor.

Constructors

RegisterCACertificateResponse'

Fields

certificateArn :: Maybe Text

The CA certificate ARN.

certificateId :: Maybe Text

The CA certificate identifier.

httpStatus :: Int

The response's http status code.

newRegisterCACertificateResponse

Arguments

:: Int
$sel:httpStatus:RegisterCACertificateResponse'

-> RegisterCACertificateResponse

Create a value of RegisterCACertificateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateArn:RegisterCACertificateResponse', registerCACertificateResponse_certificateArn - The CA certificate ARN.

$sel:certificateId:RegisterCACertificateResponse', registerCACertificateResponse_certificateId - The CA certificate identifier.

$sel:httpStatus:RegisterCACertificateResponse', registerCACertificateResponse_httpStatus - The response's http status code.

Response Lenses

registerCACertificateResponse_certificateArn :: Lens' RegisterCACertificateResponse (Maybe Text)

The CA certificate ARN.

registerCACertificateResponse_certificateId :: Lens' RegisterCACertificateResponse (Maybe Text)

The CA certificate identifier.

registerCACertificateResponse_httpStatus :: Lens' RegisterCACertificateResponse Int

The response's http status code.