Network.AWS.IoT.RegisterThing

Description

Provisions a thing in the device registry. RegisterThing calls other IoT control plane APIs. These calls might exceed your account level IoT Throttling Limits and cause throttle errors. Please contact Amazon Web Services Customer Support to raise your throttling limits if necessary.

Requires permission to access the RegisterThing action.

Creating a Request

data RegisterThing

See: newRegisterThing smart constructor.

Constructors

RegisterThing'

Fields

parameters :: Maybe (HashMap Text Text)

The parameters for provisioning a thing. See Provisioning Templates for more information.

templateBody :: Text

The provisioning template. See Provisioning Devices That Have Device Certificates for more information.

newRegisterThing

Arguments

:: Text
$sel:templateBody:RegisterThing'

-> RegisterThing

Create a value of RegisterThing with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameters:RegisterThing', registerThing_parameters - The parameters for provisioning a thing. See Provisioning Templates for more information.

$sel:templateBody:RegisterThing', registerThing_templateBody - The provisioning template. See Provisioning Devices That Have Device Certificates for more information.

Request Lenses

registerThing_parameters :: Lens' RegisterThing (Maybe (HashMap Text Text))

The parameters for provisioning a thing. See Provisioning Templates for more information.

registerThing_templateBody :: Lens' RegisterThing Text

The provisioning template. See Provisioning Devices That Have Device Certificates for more information.

Destructuring the Response

data RegisterThingResponse

See: newRegisterThingResponse smart constructor.

Constructors

RegisterThingResponse'

Fields

resourceArns :: Maybe (HashMap Text Text)

ARNs for the generated resources.

certificatePem :: Maybe Text

The certificate data, in PEM format.

httpStatus :: Int

The response's http status code.

newRegisterThingResponse

Arguments

:: Int
$sel:httpStatus:RegisterThingResponse'

-> RegisterThingResponse

Create a value of RegisterThingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArns:RegisterThingResponse', registerThingResponse_resourceArns - ARNs for the generated resources.

$sel:certificatePem:RegisterThingResponse', registerThingResponse_certificatePem - The certificate data, in PEM format.

$sel:httpStatus:RegisterThingResponse', registerThingResponse_httpStatus - The response's http status code.

Response Lenses

registerThingResponse_resourceArns :: Lens' RegisterThingResponse (Maybe (HashMap Text Text))

ARNs for the generated resources.

registerThingResponse_certificatePem :: Lens' RegisterThingResponse (Maybe Text)

The certificate data, in PEM format.

registerThingResponse_httpStatus :: Lens' RegisterThingResponse Int

The response's http status code.