Network.AWS.IoT.RejectCertificateTransfer

Description

Rejects a pending certificate transfer. After IoT rejects a certificate transfer, the certificate status changes from PENDING_TRANSFER to INACTIVE.

To check for pending certificate transfers, call ListCertificates to enumerate your certificates.

This operation can only be called by the transfer destination. After it is called, the certificate will be returned to the source's account in the INACTIVE state.

Requires permission to access the RejectCertificateTransfer action.

Creating a Request

data RejectCertificateTransfer

The input for the RejectCertificateTransfer operation.

See: newRejectCertificateTransfer smart constructor.

Constructors

RejectCertificateTransfer'

Fields

rejectReason :: Maybe Text

The reason the certificate transfer was rejected.

certificateId :: Text

The ID of the certificate. (The last part of the certificate ARN contains the certificate ID.)

newRejectCertificateTransfer

Arguments

:: Text
$sel:certificateId:RejectCertificateTransfer'

-> RejectCertificateTransfer

Create a value of RejectCertificateTransfer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rejectReason:RejectCertificateTransfer', rejectCertificateTransfer_rejectReason - The reason the certificate transfer was rejected.

$sel:certificateId:RejectCertificateTransfer', rejectCertificateTransfer_certificateId - The ID of the certificate. (The last part of the certificate ARN contains the certificate ID.)

Request Lenses

rejectCertificateTransfer_rejectReason :: Lens' RejectCertificateTransfer (Maybe Text)

The reason the certificate transfer was rejected.

rejectCertificateTransfer_certificateId :: Lens' RejectCertificateTransfer Text

The ID of the certificate. (The last part of the certificate ARN contains the certificate ID.)

Destructuring the Response

data RejectCertificateTransferResponse

See: newRejectCertificateTransferResponse smart constructor.

Constructors

RejectCertificateTransferResponse'

newRejectCertificateTransferResponse :: RejectCertificateTransferResponse

Create a value of RejectCertificateTransferResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.