Network.AWS.IoT.RemoveThingFromBillingGroup

Description

Removes the given thing from the billing group.

Requires permission to access the RemoveThingFromBillingGroup action.

Creating a Request

data RemoveThingFromBillingGroup

See: newRemoveThingFromBillingGroup smart constructor.

Constructors

RemoveThingFromBillingGroup'

Fields

thingArn :: Maybe Text

The ARN of the thing to be removed from the billing group.

thingName :: Maybe Text

The name of the thing to be removed from the billing group.

billingGroupArn :: Maybe Text

The ARN of the billing group.

billingGroupName :: Maybe Text

The name of the billing group.

newRemoveThingFromBillingGroup :: RemoveThingFromBillingGroup

Create a value of RemoveThingFromBillingGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thingArn:RemoveThingFromBillingGroup', removeThingFromBillingGroup_thingArn - The ARN of the thing to be removed from the billing group.

$sel:thingName:RemoveThingFromBillingGroup', removeThingFromBillingGroup_thingName - The name of the thing to be removed from the billing group.

$sel:billingGroupArn:RemoveThingFromBillingGroup', removeThingFromBillingGroup_billingGroupArn - The ARN of the billing group.

$sel:billingGroupName:RemoveThingFromBillingGroup', removeThingFromBillingGroup_billingGroupName - The name of the billing group.

Request Lenses

removeThingFromBillingGroup_thingArn :: Lens' RemoveThingFromBillingGroup (Maybe Text)

The ARN of the thing to be removed from the billing group.

removeThingFromBillingGroup_thingName :: Lens' RemoveThingFromBillingGroup (Maybe Text)

The name of the thing to be removed from the billing group.

removeThingFromBillingGroup_billingGroupArn :: Lens' RemoveThingFromBillingGroup (Maybe Text)

The ARN of the billing group.

removeThingFromBillingGroup_billingGroupName :: Lens' RemoveThingFromBillingGroup (Maybe Text)

The name of the billing group.

Destructuring the Response

data RemoveThingFromBillingGroupResponse

See: newRemoveThingFromBillingGroupResponse smart constructor.

Constructors

RemoveThingFromBillingGroupResponse'

Fields

httpStatus :: Int

The response's http status code.

newRemoveThingFromBillingGroupResponse

Arguments

:: Int
$sel:httpStatus:RemoveThingFromBillingGroupResponse'

-> RemoveThingFromBillingGroupResponse

Create a value of RemoveThingFromBillingGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:RemoveThingFromBillingGroupResponse', removeThingFromBillingGroupResponse_httpStatus - The response's http status code.

Response Lenses

removeThingFromBillingGroupResponse_httpStatus :: Lens' RemoveThingFromBillingGroupResponse Int

The response's http status code.