Network.AWS.IoT.TestAuthorization

Description

Tests if a specified principal is authorized to perform an IoT action on a specified resource. Use this to test and debug the authorization behavior of devices that connect to the IoT device gateway.

Requires permission to access the TestAuthorization action.

Creating a Request

data TestAuthorization

See: newTestAuthorization smart constructor.

Constructors

TestAuthorization'

Fields

clientId :: Maybe Text

The MQTT client ID.

policyNamesToAdd :: Maybe [Text]

When testing custom authorization, the policies specified here are treated as if they are attached to the principal being authorized.

principal :: Maybe Text

The principal. Valid principals are CertificateArn (arn:aws:iot:region:accountId:cert/certificateId), thingGroupArn (arn:aws:iot:region:accountId:thinggroup/groupName) and CognitoId (region:id).

cognitoIdentityPoolId :: Maybe Text

The Cognito identity pool ID.

policyNamesToSkip :: Maybe [Text]

When testing custom authorization, the policies specified here are treated as if they are not attached to the principal being authorized.

authInfos :: NonEmpty AuthInfo

A list of authorization info objects. Simulating authorization will create a response for each authInfo object in the list.

newTestAuthorization

Arguments

:: NonEmpty AuthInfo
$sel:authInfos:TestAuthorization'

-> TestAuthorization

Create a value of TestAuthorization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientId:TestAuthorization', testAuthorization_clientId - The MQTT client ID.

$sel:policyNamesToAdd:TestAuthorization', testAuthorization_policyNamesToAdd - When testing custom authorization, the policies specified here are treated as if they are attached to the principal being authorized.

$sel:principal:TestAuthorization', testAuthorization_principal - The principal. Valid principals are CertificateArn (arn:aws:iot:region:accountId:cert/certificateId), thingGroupArn (arn:aws:iot:region:accountId:thinggroup/groupName) and CognitoId (region:id).

$sel:cognitoIdentityPoolId:TestAuthorization', testAuthorization_cognitoIdentityPoolId - The Cognito identity pool ID.

$sel:policyNamesToSkip:TestAuthorization', testAuthorization_policyNamesToSkip - When testing custom authorization, the policies specified here are treated as if they are not attached to the principal being authorized.

$sel:authInfos:TestAuthorization', testAuthorization_authInfos - A list of authorization info objects. Simulating authorization will create a response for each authInfo object in the list.

Request Lenses

testAuthorization_clientId :: Lens' TestAuthorization (Maybe Text)

The MQTT client ID.

testAuthorization_policyNamesToAdd :: Lens' TestAuthorization (Maybe [Text])

When testing custom authorization, the policies specified here are treated as if they are attached to the principal being authorized.

testAuthorization_principal :: Lens' TestAuthorization (Maybe Text)

The principal. Valid principals are CertificateArn (arn:aws:iot:region:accountId:cert/certificateId), thingGroupArn (arn:aws:iot:region:accountId:thinggroup/groupName) and CognitoId (region:id).

testAuthorization_cognitoIdentityPoolId :: Lens' TestAuthorization (Maybe Text)

The Cognito identity pool ID.

testAuthorization_policyNamesToSkip :: Lens' TestAuthorization (Maybe [Text])

When testing custom authorization, the policies specified here are treated as if they are not attached to the principal being authorized.

testAuthorization_authInfos :: Lens' TestAuthorization (NonEmpty AuthInfo)

A list of authorization info objects. Simulating authorization will create a response for each authInfo object in the list.

Destructuring the Response

data TestAuthorizationResponse

See: newTestAuthorizationResponse smart constructor.

Constructors

TestAuthorizationResponse'

Fields

authResults :: Maybe [AuthResult]

The authentication results.

httpStatus :: Int

The response's http status code.

newTestAuthorizationResponse

Arguments

:: Int
$sel:httpStatus:TestAuthorizationResponse'

-> TestAuthorizationResponse

Create a value of TestAuthorizationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authResults:TestAuthorizationResponse', testAuthorizationResponse_authResults - The authentication results.

$sel:httpStatus:TestAuthorizationResponse', testAuthorizationResponse_httpStatus - The response's http status code.

Response Lenses

testAuthorizationResponse_authResults :: Lens' TestAuthorizationResponse (Maybe [AuthResult])

The authentication results.

testAuthorizationResponse_httpStatus :: Lens' TestAuthorizationResponse Int

The response's http status code.