Network.AWS.IoT.Types.Action

Description

Documentation

data Action

Describes the actions associated with a rule.

See: newAction smart constructor.

Constructors

Action'

Fields

cloudwatchMetric :: Maybe CloudwatchMetricAction

Capture a CloudWatch metric.

cloudwatchLogs :: Maybe CloudwatchLogsAction

Send data to CloudWatch Logs.

dynamoDBv2 :: Maybe DynamoDBv2Action

Write to a DynamoDB table. This is a new version of the DynamoDB action. It allows you to write each attribute in an MQTT message payload into a separate DynamoDB column.

stepFunctions :: Maybe StepFunctionsAction

Starts execution of a Step Functions state machine.

cloudwatchAlarm :: Maybe CloudwatchAlarmAction

Change the state of a CloudWatch alarm.

sns :: Maybe SnsAction

Publish to an Amazon SNS topic.

dynamoDB :: Maybe DynamoDBAction

Write to a DynamoDB table.

firehose :: Maybe FirehoseAction

Write to an Amazon Kinesis Firehose stream.

timestream :: Maybe TimestreamAction

The Timestream rule action writes attributes (measures) from an MQTT message into an Amazon Timestream table. For more information, see the Timestream topic rule action documentation.

iotSiteWise :: Maybe IotSiteWiseAction

Sends data from the MQTT message that triggered the rule to IoT SiteWise asset properties.

iotAnalytics :: Maybe IotAnalyticsAction

Sends message data to an IoT Analytics channel.

lambda :: Maybe LambdaAction

Invoke a Lambda function.

openSearch :: Maybe OpenSearchAction

Write data to an Amazon OpenSearch Service domain.

iotEvents :: Maybe IotEventsAction

Sends an input to an IoT Events detector.

salesforce :: Maybe SalesforceAction

Send a message to a Salesforce IoT Cloud Input Stream.

kinesis :: Maybe KinesisAction

Write data to an Amazon Kinesis stream.

s3 :: Maybe S3Action

Write to an Amazon S3 bucket.

http :: Maybe HttpAction

Send data to an HTTPS endpoint.

elasticsearch :: Maybe ElasticsearchAction

Write data to an Amazon OpenSearch Service domain.

The Elasticsearch action can only be used by existing rule actions. To create a new rule action or to update an existing rule action, use the OpenSearch rule action instead. For more information, see OpenSearchAction.

kafka :: Maybe KafkaAction

Send messages to an Amazon Managed Streaming for Apache Kafka (Amazon MSK) or self-managed Apache Kafka cluster.

republish :: Maybe RepublishAction

Publish to another MQTT topic.

sqs :: Maybe SqsAction

Publish to an Amazon SQS queue.

newAction :: Action

Create a value of Action with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cloudwatchMetric:Action', action_cloudwatchMetric - Capture a CloudWatch metric.

$sel:cloudwatchLogs:Action', action_cloudwatchLogs - Send data to CloudWatch Logs.

$sel:dynamoDBv2:Action', action_dynamoDBv2 - Write to a DynamoDB table. This is a new version of the DynamoDB action. It allows you to write each attribute in an MQTT message payload into a separate DynamoDB column.

$sel:stepFunctions:Action', action_stepFunctions - Starts execution of a Step Functions state machine.

$sel:cloudwatchAlarm:Action', action_cloudwatchAlarm - Change the state of a CloudWatch alarm.

$sel:sns:Action', action_sns - Publish to an Amazon SNS topic.

$sel:dynamoDB:Action', action_dynamoDB - Write to a DynamoDB table.

$sel:firehose:Action', action_firehose - Write to an Amazon Kinesis Firehose stream.

$sel:timestream:Action', action_timestream - The Timestream rule action writes attributes (measures) from an MQTT message into an Amazon Timestream table. For more information, see the Timestream topic rule action documentation.

$sel:iotSiteWise:Action', action_iotSiteWise - Sends data from the MQTT message that triggered the rule to IoT SiteWise asset properties.

$sel:iotAnalytics:Action', action_iotAnalytics - Sends message data to an IoT Analytics channel.

$sel:lambda:Action', action_lambda - Invoke a Lambda function.

$sel:openSearch:Action', action_openSearch - Write data to an Amazon OpenSearch Service domain.

$sel:iotEvents:Action', action_iotEvents - Sends an input to an IoT Events detector.

$sel:salesforce:Action', action_salesforce - Send a message to a Salesforce IoT Cloud Input Stream.

$sel:kinesis:Action', action_kinesis - Write data to an Amazon Kinesis stream.

$sel:s3:Action', action_s3 - Write to an Amazon S3 bucket.

$sel:http:Action', action_http - Send data to an HTTPS endpoint.

$sel:elasticsearch:Action', action_elasticsearch - Write data to an Amazon OpenSearch Service domain.

The Elasticsearch action can only be used by existing rule actions. To create a new rule action or to update an existing rule action, use the OpenSearch rule action instead. For more information, see OpenSearchAction.

$sel:kafka:Action', action_kafka - Send messages to an Amazon Managed Streaming for Apache Kafka (Amazon MSK) or self-managed Apache Kafka cluster.

$sel:republish:Action', action_republish - Publish to another MQTT topic.

$sel:sqs:Action', action_sqs - Publish to an Amazon SQS queue.

action_cloudwatchMetric :: Lens' Action (Maybe CloudwatchMetricAction)

Capture a CloudWatch metric.

action_cloudwatchLogs :: Lens' Action (Maybe CloudwatchLogsAction)

Send data to CloudWatch Logs.

action_dynamoDBv2 :: Lens' Action (Maybe DynamoDBv2Action)

Write to a DynamoDB table. This is a new version of the DynamoDB action. It allows you to write each attribute in an MQTT message payload into a separate DynamoDB column.

action_stepFunctions :: Lens' Action (Maybe StepFunctionsAction)

Starts execution of a Step Functions state machine.

action_cloudwatchAlarm :: Lens' Action (Maybe CloudwatchAlarmAction)

Change the state of a CloudWatch alarm.

action_sns :: Lens' Action (Maybe SnsAction)

Publish to an Amazon SNS topic.

action_dynamoDB :: Lens' Action (Maybe DynamoDBAction)

Write to a DynamoDB table.

action_firehose :: Lens' Action (Maybe FirehoseAction)

Write to an Amazon Kinesis Firehose stream.

action_timestream :: Lens' Action (Maybe TimestreamAction)

The Timestream rule action writes attributes (measures) from an MQTT message into an Amazon Timestream table. For more information, see the Timestream topic rule action documentation.

action_iotSiteWise :: Lens' Action (Maybe IotSiteWiseAction)

Sends data from the MQTT message that triggered the rule to IoT SiteWise asset properties.

action_iotAnalytics :: Lens' Action (Maybe IotAnalyticsAction)

Sends message data to an IoT Analytics channel.

action_lambda :: Lens' Action (Maybe LambdaAction)

Invoke a Lambda function.

action_openSearch :: Lens' Action (Maybe OpenSearchAction)

Write data to an Amazon OpenSearch Service domain.

action_iotEvents :: Lens' Action (Maybe IotEventsAction)

Sends an input to an IoT Events detector.

action_salesforce :: Lens' Action (Maybe SalesforceAction)

Send a message to a Salesforce IoT Cloud Input Stream.

action_kinesis :: Lens' Action (Maybe KinesisAction)

Write data to an Amazon Kinesis stream.

action_s3 :: Lens' Action (Maybe S3Action)

Write to an Amazon S3 bucket.

action_http :: Lens' Action (Maybe HttpAction)

Send data to an HTTPS endpoint.

action_elasticsearch :: Lens' Action (Maybe ElasticsearchAction)

Write data to an Amazon OpenSearch Service domain.

The Elasticsearch action can only be used by existing rule actions. To create a new rule action or to update an existing rule action, use the OpenSearch rule action instead. For more information, see OpenSearchAction.

action_kafka :: Lens' Action (Maybe KafkaAction)

Send messages to an Amazon Managed Streaming for Apache Kafka (Amazon MSK) or self-managed Apache Kafka cluster.

action_republish :: Lens' Action (Maybe RepublishAction)

Publish to another MQTT topic.

action_sqs :: Lens' Action (Maybe SqsAction)

Publish to an Amazon SQS queue.