Network.AWS.IoT.Types.ActiveViolation

Description

Documentation

data ActiveViolation

Information about an active Device Defender security profile behavior violation.

See: newActiveViolation smart constructor.

Constructors

ActiveViolation'

Fields

violationId :: Maybe Text

The ID of the active violation.

lastViolationTime :: Maybe POSIX

The time the most recent violation occurred.

thingName :: Maybe Text

The name of the thing responsible for the active violation.

lastViolationValue :: Maybe MetricValue

The value of the metric (the measurement) that caused the most recent violation.

verificationStateDescription :: Maybe Text

The description of the verification state of the violation.

securityProfileName :: Maybe Text

The security profile with the behavior is in violation.

behavior :: Maybe Behavior

The behavior that is being violated.

violationStartTime :: Maybe POSIX

The time the violation started.

verificationState :: Maybe VerificationState

The verification state of the violation (detect alarm).

violationEventAdditionalInfo :: Maybe ViolationEventAdditionalInfo

The details of a violation event.

newActiveViolation :: ActiveViolation

Create a value of ActiveViolation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:violationId:ActiveViolation', activeViolation_violationId - The ID of the active violation.

$sel:lastViolationTime:ActiveViolation', activeViolation_lastViolationTime - The time the most recent violation occurred.

$sel:thingName:ActiveViolation', activeViolation_thingName - The name of the thing responsible for the active violation.

$sel:lastViolationValue:ActiveViolation', activeViolation_lastViolationValue - The value of the metric (the measurement) that caused the most recent violation.

$sel:verificationStateDescription:ActiveViolation', activeViolation_verificationStateDescription - The description of the verification state of the violation.

$sel:securityProfileName:ActiveViolation', activeViolation_securityProfileName - The security profile with the behavior is in violation.

$sel:behavior:ActiveViolation', activeViolation_behavior - The behavior that is being violated.

$sel:violationStartTime:ActiveViolation', activeViolation_violationStartTime - The time the violation started.

$sel:verificationState:ActiveViolation', activeViolation_verificationState - The verification state of the violation (detect alarm).

$sel:violationEventAdditionalInfo:ActiveViolation', activeViolation_violationEventAdditionalInfo - The details of a violation event.

activeViolation_violationId :: Lens' ActiveViolation (Maybe Text)

The ID of the active violation.

activeViolation_lastViolationTime :: Lens' ActiveViolation (Maybe UTCTime)

The time the most recent violation occurred.

activeViolation_thingName :: Lens' ActiveViolation (Maybe Text)

The name of the thing responsible for the active violation.

activeViolation_lastViolationValue :: Lens' ActiveViolation (Maybe MetricValue)

The value of the metric (the measurement) that caused the most recent violation.

activeViolation_verificationStateDescription :: Lens' ActiveViolation (Maybe Text)

The description of the verification state of the violation.

activeViolation_securityProfileName :: Lens' ActiveViolation (Maybe Text)

The security profile with the behavior is in violation.

activeViolation_behavior :: Lens' ActiveViolation (Maybe Behavior)

The behavior that is being violated.

activeViolation_violationStartTime :: Lens' ActiveViolation (Maybe UTCTime)

The time the violation started.

activeViolation_verificationState :: Lens' ActiveViolation (Maybe VerificationState)

The verification state of the violation (detect alarm).

activeViolation_violationEventAdditionalInfo :: Lens' ActiveViolation (Maybe ViolationEventAdditionalInfo)

The details of a violation event.