Network.AWS.IoT.Types.AuditCheckDetails

Description

Documentation

data AuditCheckDetails

Information about the audit check.

See: newAuditCheckDetails smart constructor.

Constructors

AuditCheckDetails'

Fields

checkCompliant :: Maybe Bool

True if the check is complete and found all resources compliant.

message :: Maybe Text

The message associated with any error encountered when this check is performed during this audit.

suppressedNonCompliantResourcesCount :: Maybe Integer

Describes how many of the non-compliant resources created during the evaluation of an audit check were marked as suppressed.

checkRunStatus :: Maybe AuditCheckRunStatus

The completion status of this check. One of "IN_PROGRESS", "WAITING_FOR_DATA_COLLECTION", "CANCELED", "COMPLETED_COMPLIANT", "COMPLETED_NON_COMPLIANT", or "FAILED".

totalResourcesCount :: Maybe Integer

The number of resources on which the check was performed.

errorCode :: Maybe Text

The code of any error encountered when this check is performed during this audit. One of "INSUFFICIENT_PERMISSIONS" or "AUDIT_CHECK_DISABLED".

nonCompliantResourcesCount :: Maybe Integer

The number of resources that were found noncompliant during the check.

newAuditCheckDetails :: AuditCheckDetails

Create a value of AuditCheckDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checkCompliant:AuditCheckDetails', auditCheckDetails_checkCompliant - True if the check is complete and found all resources compliant.

$sel:message:AuditCheckDetails', auditCheckDetails_message - The message associated with any error encountered when this check is performed during this audit.

$sel:suppressedNonCompliantResourcesCount:AuditCheckDetails', auditCheckDetails_suppressedNonCompliantResourcesCount - Describes how many of the non-compliant resources created during the evaluation of an audit check were marked as suppressed.

$sel:checkRunStatus:AuditCheckDetails', auditCheckDetails_checkRunStatus - The completion status of this check. One of "IN_PROGRESS", "WAITING_FOR_DATA_COLLECTION", "CANCELED", "COMPLETED_COMPLIANT", "COMPLETED_NON_COMPLIANT", or "FAILED".

$sel:totalResourcesCount:AuditCheckDetails', auditCheckDetails_totalResourcesCount - The number of resources on which the check was performed.

$sel:errorCode:AuditCheckDetails', auditCheckDetails_errorCode - The code of any error encountered when this check is performed during this audit. One of "INSUFFICIENT_PERMISSIONS" or "AUDIT_CHECK_DISABLED".

$sel:nonCompliantResourcesCount:AuditCheckDetails', auditCheckDetails_nonCompliantResourcesCount - The number of resources that were found noncompliant during the check.

auditCheckDetails_checkCompliant :: Lens' AuditCheckDetails (Maybe Bool)

True if the check is complete and found all resources compliant.

auditCheckDetails_message :: Lens' AuditCheckDetails (Maybe Text)

The message associated with any error encountered when this check is performed during this audit.

auditCheckDetails_suppressedNonCompliantResourcesCount :: Lens' AuditCheckDetails (Maybe Integer)

Describes how many of the non-compliant resources created during the evaluation of an audit check were marked as suppressed.

auditCheckDetails_checkRunStatus :: Lens' AuditCheckDetails (Maybe AuditCheckRunStatus)

The completion status of this check. One of "IN_PROGRESS", "WAITING_FOR_DATA_COLLECTION", "CANCELED", "COMPLETED_COMPLIANT", "COMPLETED_NON_COMPLIANT", or "FAILED".

auditCheckDetails_totalResourcesCount :: Lens' AuditCheckDetails (Maybe Integer)

The number of resources on which the check was performed.

auditCheckDetails_errorCode :: Lens' AuditCheckDetails (Maybe Text)

The code of any error encountered when this check is performed during this audit. One of "INSUFFICIENT_PERMISSIONS" or "AUDIT_CHECK_DISABLED".

auditCheckDetails_nonCompliantResourcesCount :: Lens' AuditCheckDetails (Maybe Integer)

The number of resources that were found noncompliant during the check.