Network.AWS.IoT.Types.AuditMitigationActionExecutionMetadata

Description

Documentation

data AuditMitigationActionExecutionMetadata

Returned by ListAuditMitigationActionsTask, this object contains information that describes a mitigation action that has been started.

See: newAuditMitigationActionExecutionMetadata smart constructor.

Constructors

AuditMitigationActionExecutionMetadata'

Fields

status :: Maybe AuditMitigationActionsExecutionStatus

The current status of the task being executed.

startTime :: Maybe POSIX

The date and time when the task was started.

taskId :: Maybe Text

The unique identifier for the task that applies the mitigation action.

actionId :: Maybe Text

The unique identifier for the mitigation action being applied by the task.

actionName :: Maybe Text

The friendly name of the mitigation action being applied by the task.

endTime :: Maybe POSIX

The date and time when the task was completed or canceled. Blank if the task is still running.

errorCode :: Maybe Text

If an error occurred, the code that indicates which type of error occurred.

findingId :: Maybe Text

The unique identifier for the findings to which the task and associated mitigation action are applied.

message :: Maybe Text

If an error occurred, a message that describes the error.

newAuditMitigationActionExecutionMetadata :: AuditMitigationActionExecutionMetadata

Create a value of AuditMitigationActionExecutionMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AuditMitigationActionExecutionMetadata', auditMitigationActionExecutionMetadata_status - The current status of the task being executed.

$sel:startTime:AuditMitigationActionExecutionMetadata', auditMitigationActionExecutionMetadata_startTime - The date and time when the task was started.

$sel:taskId:AuditMitigationActionExecutionMetadata', auditMitigationActionExecutionMetadata_taskId - The unique identifier for the task that applies the mitigation action.

$sel:actionId:AuditMitigationActionExecutionMetadata', auditMitigationActionExecutionMetadata_actionId - The unique identifier for the mitigation action being applied by the task.

$sel:actionName:AuditMitigationActionExecutionMetadata', auditMitigationActionExecutionMetadata_actionName - The friendly name of the mitigation action being applied by the task.

$sel:endTime:AuditMitigationActionExecutionMetadata', auditMitigationActionExecutionMetadata_endTime - The date and time when the task was completed or canceled. Blank if the task is still running.

$sel:errorCode:AuditMitigationActionExecutionMetadata', auditMitigationActionExecutionMetadata_errorCode - If an error occurred, the code that indicates which type of error occurred.

$sel:findingId:AuditMitigationActionExecutionMetadata', auditMitigationActionExecutionMetadata_findingId - The unique identifier for the findings to which the task and associated mitigation action are applied.

$sel:message:AuditMitigationActionExecutionMetadata', auditMitigationActionExecutionMetadata_message - If an error occurred, a message that describes the error.

auditMitigationActionExecutionMetadata_status :: Lens' AuditMitigationActionExecutionMetadata (Maybe AuditMitigationActionsExecutionStatus)

The current status of the task being executed.

auditMitigationActionExecutionMetadata_startTime :: Lens' AuditMitigationActionExecutionMetadata (Maybe UTCTime)

The date and time when the task was started.

auditMitigationActionExecutionMetadata_taskId :: Lens' AuditMitigationActionExecutionMetadata (Maybe Text)

The unique identifier for the task that applies the mitigation action.

auditMitigationActionExecutionMetadata_actionId :: Lens' AuditMitigationActionExecutionMetadata (Maybe Text)

The unique identifier for the mitigation action being applied by the task.

auditMitigationActionExecutionMetadata_actionName :: Lens' AuditMitigationActionExecutionMetadata (Maybe Text)

The friendly name of the mitigation action being applied by the task.

auditMitigationActionExecutionMetadata_endTime :: Lens' AuditMitigationActionExecutionMetadata (Maybe UTCTime)

The date and time when the task was completed or canceled. Blank if the task is still running.

auditMitigationActionExecutionMetadata_errorCode :: Lens' AuditMitigationActionExecutionMetadata (Maybe Text)

If an error occurred, the code that indicates which type of error occurred.

auditMitigationActionExecutionMetadata_findingId :: Lens' AuditMitigationActionExecutionMetadata (Maybe Text)

The unique identifier for the findings to which the task and associated mitigation action are applied.

auditMitigationActionExecutionMetadata_message :: Lens' AuditMitigationActionExecutionMetadata (Maybe Text)

If an error occurred, a message that describes the error.