Network.AWS.IoT.Types.AuditSuppression

Description

Documentation

data AuditSuppression

Filters out specific findings of a Device Defender audit.

See: newAuditSuppression smart constructor.

Constructors

AuditSuppression'

Fields

expirationDate :: Maybe POSIX

The expiration date (epoch timestamp in seconds) that you want the suppression to adhere to.

suppressIndefinitely :: Maybe Bool

Indicates whether a suppression should exist indefinitely or not.

description :: Maybe Text

The description of the audit suppression.

checkName :: Text

resourceIdentifier :: ResourceIdentifier

newAuditSuppression

Arguments

:: Text
$sel:checkName:AuditSuppression'

-> ResourceIdentifier
$sel:resourceIdentifier:AuditSuppression'

-> AuditSuppression

Create a value of AuditSuppression with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expirationDate:AuditSuppression', auditSuppression_expirationDate - The expiration date (epoch timestamp in seconds) that you want the suppression to adhere to.

$sel:suppressIndefinitely:AuditSuppression', auditSuppression_suppressIndefinitely - Indicates whether a suppression should exist indefinitely or not.

$sel:description:AuditSuppression', auditSuppression_description - The description of the audit suppression.

$sel:checkName:AuditSuppression', auditSuppression_checkName - Undocumented member.

$sel:resourceIdentifier:AuditSuppression', auditSuppression_resourceIdentifier - Undocumented member.

auditSuppression_expirationDate :: Lens' AuditSuppression (Maybe UTCTime)

The expiration date (epoch timestamp in seconds) that you want the suppression to adhere to.

auditSuppression_suppressIndefinitely :: Lens' AuditSuppression (Maybe Bool)

Indicates whether a suppression should exist indefinitely or not.

auditSuppression_description :: Lens' AuditSuppression (Maybe Text)

The description of the audit suppression.

auditSuppression_checkName :: Lens' AuditSuppression Text

Undocumented member.

auditSuppression_resourceIdentifier :: Lens' AuditSuppression ResourceIdentifier

Undocumented member.