Network.AWS.IoT.Types.AuthResult

Description

Documentation

data AuthResult

The authorizer result.

See: newAuthResult smart constructor.

Constructors

AuthResult'

Fields

authInfo :: Maybe AuthInfo

Authorization information.

allowed :: Maybe Allowed

The policies and statements that allowed the specified action.

denied :: Maybe Denied

The policies and statements that denied the specified action.

missingContextValues :: Maybe [Text]

Contains any missing context values found while evaluating policy.

authDecision :: Maybe AuthDecision

The final authorization decision of this scenario. Multiple statements are taken into account when determining the authorization decision. An explicit deny statement can override multiple allow statements.

newAuthResult :: AuthResult

Create a value of AuthResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authInfo:AuthResult', authResult_authInfo - Authorization information.

$sel:allowed:AuthResult', authResult_allowed - The policies and statements that allowed the specified action.

$sel:denied:AuthResult', authResult_denied - The policies and statements that denied the specified action.

$sel:missingContextValues:AuthResult', authResult_missingContextValues - Contains any missing context values found while evaluating policy.

$sel:authDecision:AuthResult', authResult_authDecision - The final authorization decision of this scenario. Multiple statements are taken into account when determining the authorization decision. An explicit deny statement can override multiple allow statements.

authResult_authInfo :: Lens' AuthResult (Maybe AuthInfo)

Authorization information.

authResult_allowed :: Lens' AuthResult (Maybe Allowed)

The policies and statements that allowed the specified action.

authResult_denied :: Lens' AuthResult (Maybe Denied)

The policies and statements that denied the specified action.

authResult_missingContextValues :: Lens' AuthResult (Maybe [Text])

Contains any missing context values found while evaluating policy.

authResult_authDecision :: Lens' AuthResult (Maybe AuthDecision)

The final authorization decision of this scenario. Multiple statements are taken into account when determining the authorization decision. An explicit deny statement can override multiple allow statements.