Network.AWS.IoT.Types.AuthorizerDescription

Description

Documentation

data AuthorizerDescription

The authorizer description.

See: newAuthorizerDescription smart constructor.

Constructors

AuthorizerDescription'

Fields

lastModifiedDate :: Maybe POSIX

The UNIX timestamp of when the authorizer was last updated.

authorizerArn :: Maybe Text

The authorizer ARN.

status :: Maybe AuthorizerStatus

The status of the authorizer.

authorizerFunctionArn :: Maybe Text

The authorizer's Lambda function ARN.

creationDate :: Maybe POSIX

The UNIX timestamp of when the authorizer was created.

tokenSigningPublicKeys :: Maybe (HashMap Text Text)

The public keys used to validate the token signature returned by your custom authentication service.

authorizerName :: Maybe Text

The authorizer name.

tokenKeyName :: Maybe Text

The key used to extract the token from the HTTP headers.

signingDisabled :: Maybe Bool

Specifies whether IoT validates the token signature in an authorization request.

newAuthorizerDescription :: AuthorizerDescription

Create a value of AuthorizerDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:AuthorizerDescription', authorizerDescription_lastModifiedDate - The UNIX timestamp of when the authorizer was last updated.

$sel:authorizerArn:AuthorizerDescription', authorizerDescription_authorizerArn - The authorizer ARN.

$sel:status:AuthorizerDescription', authorizerDescription_status - The status of the authorizer.

$sel:authorizerFunctionArn:AuthorizerDescription', authorizerDescription_authorizerFunctionArn - The authorizer's Lambda function ARN.

$sel:creationDate:AuthorizerDescription', authorizerDescription_creationDate - The UNIX timestamp of when the authorizer was created.

$sel:tokenSigningPublicKeys:AuthorizerDescription', authorizerDescription_tokenSigningPublicKeys - The public keys used to validate the token signature returned by your custom authentication service.

$sel:authorizerName:AuthorizerDescription', authorizerDescription_authorizerName - The authorizer name.

$sel:tokenKeyName:AuthorizerDescription', authorizerDescription_tokenKeyName - The key used to extract the token from the HTTP headers.

$sel:signingDisabled:AuthorizerDescription', authorizerDescription_signingDisabled - Specifies whether IoT validates the token signature in an authorization request.

authorizerDescription_lastModifiedDate :: Lens' AuthorizerDescription (Maybe UTCTime)

The UNIX timestamp of when the authorizer was last updated.

authorizerDescription_authorizerArn :: Lens' AuthorizerDescription (Maybe Text)

The authorizer ARN.

authorizerDescription_status :: Lens' AuthorizerDescription (Maybe AuthorizerStatus)

The status of the authorizer.

authorizerDescription_authorizerFunctionArn :: Lens' AuthorizerDescription (Maybe Text)

The authorizer's Lambda function ARN.

authorizerDescription_creationDate :: Lens' AuthorizerDescription (Maybe UTCTime)

The UNIX timestamp of when the authorizer was created.

authorizerDescription_tokenSigningPublicKeys :: Lens' AuthorizerDescription (Maybe (HashMap Text Text))

The public keys used to validate the token signature returned by your custom authentication service.

authorizerDescription_authorizerName :: Lens' AuthorizerDescription (Maybe Text)

The authorizer name.

authorizerDescription_tokenKeyName :: Lens' AuthorizerDescription (Maybe Text)

The key used to extract the token from the HTTP headers.

authorizerDescription_signingDisabled :: Lens' AuthorizerDescription (Maybe Bool)

Specifies whether IoT validates the token signature in an authorization request.