Network.AWS.IoT.Types.AwsJobExponentialRolloutRate

Description

Documentation

data AwsJobExponentialRolloutRate

The rate of increase for a job rollout. This parameter allows you to define an exponential rate increase for a job rollout.

See: newAwsJobExponentialRolloutRate smart constructor.

Constructors

AwsJobExponentialRolloutRate'

Fields

baseRatePerMinute :: Natural

The minimum number of things that will be notified of a pending job, per minute, at the start of the job rollout. This is the initial rate of the rollout.

incrementFactor :: Double

The rate of increase for a job rollout. The number of things notified is multiplied by this factor.

rateIncreaseCriteria :: AwsJobRateIncreaseCriteria

The criteria to initiate the increase in rate of rollout for a job.

Amazon Web Services IoT Core supports up to one digit after the decimal (for example, 1.5, but not 1.55).

newAwsJobExponentialRolloutRate

Arguments

:: Natural
$sel:baseRatePerMinute:AwsJobExponentialRolloutRate'

-> Double
$sel:incrementFactor:AwsJobExponentialRolloutRate'

-> AwsJobRateIncreaseCriteria
$sel:rateIncreaseCriteria:AwsJobExponentialRolloutRate'

-> AwsJobExponentialRolloutRate

Create a value of AwsJobExponentialRolloutRate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:baseRatePerMinute:AwsJobExponentialRolloutRate', awsJobExponentialRolloutRate_baseRatePerMinute - The minimum number of things that will be notified of a pending job, per minute, at the start of the job rollout. This is the initial rate of the rollout.

$sel:incrementFactor:AwsJobExponentialRolloutRate', awsJobExponentialRolloutRate_incrementFactor - The rate of increase for a job rollout. The number of things notified is multiplied by this factor.

$sel:rateIncreaseCriteria:AwsJobExponentialRolloutRate', awsJobExponentialRolloutRate_rateIncreaseCriteria - The criteria to initiate the increase in rate of rollout for a job.

Amazon Web Services IoT Core supports up to one digit after the decimal (for example, 1.5, but not 1.55).

awsJobExponentialRolloutRate_baseRatePerMinute :: Lens' AwsJobExponentialRolloutRate Natural

The minimum number of things that will be notified of a pending job, per minute, at the start of the job rollout. This is the initial rate of the rollout.

awsJobExponentialRolloutRate_incrementFactor :: Lens' AwsJobExponentialRolloutRate Double

The rate of increase for a job rollout. The number of things notified is multiplied by this factor.

awsJobExponentialRolloutRate_rateIncreaseCriteria :: Lens' AwsJobExponentialRolloutRate AwsJobRateIncreaseCriteria

The criteria to initiate the increase in rate of rollout for a job.

Amazon Web Services IoT Core supports up to one digit after the decimal (for example, 1.5, but not 1.55).