Network.AWS.IoT.Types.BehaviorCriteria

Description

Documentation

data BehaviorCriteria

The criteria by which the behavior is determined to be normal.

See: newBehaviorCriteria smart constructor.

Constructors

BehaviorCriteria'

Fields

comparisonOperator :: Maybe ComparisonOperator

The operator that relates the thing measured (metric) to the criteria (containing a value or statisticalThreshold). Valid operators include:

string-list: in-set and not-in-set

number-list: in-set and not-in-set

ip-address-list: in-cidr-set and not-in-cidr-set

number: less-than, less-than-equals, greater-than, and greater-than-equals

consecutiveDatapointsToAlarm :: Maybe Natural

If a device is in violation of the behavior for the specified number of consecutive datapoints, an alarm occurs. If not specified, the default is 1.

mlDetectionConfig :: Maybe MachineLearningDetectionConfig

The configuration of an ML Detect

statisticalThreshold :: Maybe StatisticalThreshold

A statistical ranking (percentile)that indicates a threshold value by which a behavior is determined to be in compliance or in violation of the behavior.

consecutiveDatapointsToClear :: Maybe Natural

If an alarm has occurred and the offending device is no longer in violation of the behavior for the specified number of consecutive datapoints, the alarm is cleared. If not specified, the default is 1.

value :: Maybe MetricValue

The value to be compared with the metric.

durationSeconds :: Maybe Int

Use this to specify the time duration over which the behavior is evaluated, for those criteria that have a time dimension (for example, NUM_MESSAGES_SENT). For a statisticalThreshhold metric comparison, measurements from all devices are accumulated over this time duration before being used to calculate percentiles, and later, measurements from an individual device are also accumulated over this time duration before being given a percentile rank. Cannot be used with list-based metric datatypes.

newBehaviorCriteria :: BehaviorCriteria

Create a value of BehaviorCriteria with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:comparisonOperator:BehaviorCriteria', behaviorCriteria_comparisonOperator - The operator that relates the thing measured (metric) to the criteria (containing a value or statisticalThreshold). Valid operators include:

string-list: in-set and not-in-set

number-list: in-set and not-in-set

ip-address-list: in-cidr-set and not-in-cidr-set

number: less-than, less-than-equals, greater-than, and greater-than-equals

$sel:consecutiveDatapointsToAlarm:BehaviorCriteria', behaviorCriteria_consecutiveDatapointsToAlarm - If a device is in violation of the behavior for the specified number of consecutive datapoints, an alarm occurs. If not specified, the default is 1.

$sel:mlDetectionConfig:BehaviorCriteria', behaviorCriteria_mlDetectionConfig - The configuration of an ML Detect

$sel:statisticalThreshold:BehaviorCriteria', behaviorCriteria_statisticalThreshold - A statistical ranking (percentile)that indicates a threshold value by which a behavior is determined to be in compliance or in violation of the behavior.

$sel:consecutiveDatapointsToClear:BehaviorCriteria', behaviorCriteria_consecutiveDatapointsToClear - If an alarm has occurred and the offending device is no longer in violation of the behavior for the specified number of consecutive datapoints, the alarm is cleared. If not specified, the default is 1.

$sel:value:BehaviorCriteria', behaviorCriteria_value - The value to be compared with the metric.

$sel:durationSeconds:BehaviorCriteria', behaviorCriteria_durationSeconds - Use this to specify the time duration over which the behavior is evaluated, for those criteria that have a time dimension (for example, NUM_MESSAGES_SENT). For a statisticalThreshhold metric comparison, measurements from all devices are accumulated over this time duration before being used to calculate percentiles, and later, measurements from an individual device are also accumulated over this time duration before being given a percentile rank. Cannot be used with list-based metric datatypes.

behaviorCriteria_comparisonOperator :: Lens' BehaviorCriteria (Maybe ComparisonOperator)

The operator that relates the thing measured (metric) to the criteria (containing a value or statisticalThreshold). Valid operators include:

string-list: in-set and not-in-set

number-list: in-set and not-in-set

ip-address-list: in-cidr-set and not-in-cidr-set

number: less-than, less-than-equals, greater-than, and greater-than-equals

behaviorCriteria_consecutiveDatapointsToAlarm :: Lens' BehaviorCriteria (Maybe Natural)

If a device is in violation of the behavior for the specified number of consecutive datapoints, an alarm occurs. If not specified, the default is 1.

behaviorCriteria_mlDetectionConfig :: Lens' BehaviorCriteria (Maybe MachineLearningDetectionConfig)

The configuration of an ML Detect

behaviorCriteria_statisticalThreshold :: Lens' BehaviorCriteria (Maybe StatisticalThreshold)

A statistical ranking (percentile)that indicates a threshold value by which a behavior is determined to be in compliance or in violation of the behavior.

behaviorCriteria_consecutiveDatapointsToClear :: Lens' BehaviorCriteria (Maybe Natural)

If an alarm has occurred and the offending device is no longer in violation of the behavior for the specified number of consecutive datapoints, the alarm is cleared. If not specified, the default is 1.

behaviorCriteria_value :: Lens' BehaviorCriteria (Maybe MetricValue)

The value to be compared with the metric.

behaviorCriteria_durationSeconds :: Lens' BehaviorCriteria (Maybe Int)

Use this to specify the time duration over which the behavior is evaluated, for those criteria that have a time dimension (for example, NUM_MESSAGES_SENT). For a statisticalThreshhold metric comparison, measurements from all devices are accumulated over this time duration before being used to calculate percentiles, and later, measurements from an individual device are also accumulated over this time duration before being given a percentile rank. Cannot be used with list-based metric datatypes.