Network.AWS.IoT.Types.CACertificateDescription

Description

Documentation

data CACertificateDescription

Describes a CA certificate.

See: newCACertificateDescription smart constructor.

Constructors

CACertificateDescription'

Fields

status :: Maybe CACertificateStatus

The status of a CA certificate.

ownedBy :: Maybe Text

The owner of the CA certificate.

lastModifiedDate :: Maybe POSIX

The date the CA certificate was last modified.

certificatePem :: Maybe Text

The CA certificate data, in PEM format.

certificateArn :: Maybe Text

The CA certificate ARN.

certificateId :: Maybe Text

The CA certificate ID.

validity :: Maybe CertificateValidity

When the CA certificate is valid.

autoRegistrationStatus :: Maybe AutoRegistrationStatus

Whether the CA certificate configured for auto registration of device certificates. Valid values are "ENABLE" and "DISABLE"

creationDate :: Maybe POSIX

The date the CA certificate was created.

generationId :: Maybe Text

The generation ID of the CA certificate.

customerVersion :: Maybe Natural

The customer version of the CA certificate.

newCACertificateDescription :: CACertificateDescription

Create a value of CACertificateDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CACertificateDescription', cACertificateDescription_status - The status of a CA certificate.

$sel:ownedBy:CACertificateDescription', cACertificateDescription_ownedBy - The owner of the CA certificate.

$sel:lastModifiedDate:CACertificateDescription', cACertificateDescription_lastModifiedDate - The date the CA certificate was last modified.

$sel:certificatePem:CACertificateDescription', cACertificateDescription_certificatePem - The CA certificate data, in PEM format.

$sel:certificateArn:CACertificateDescription', cACertificateDescription_certificateArn - The CA certificate ARN.

$sel:certificateId:CACertificateDescription', cACertificateDescription_certificateId - The CA certificate ID.

$sel:validity:CACertificateDescription', cACertificateDescription_validity - When the CA certificate is valid.

$sel:autoRegistrationStatus:CACertificateDescription', cACertificateDescription_autoRegistrationStatus - Whether the CA certificate configured for auto registration of device certificates. Valid values are "ENABLE" and "DISABLE"

$sel:creationDate:CACertificateDescription', cACertificateDescription_creationDate - The date the CA certificate was created.

$sel:generationId:CACertificateDescription', cACertificateDescription_generationId - The generation ID of the CA certificate.

$sel:customerVersion:CACertificateDescription', cACertificateDescription_customerVersion - The customer version of the CA certificate.

cACertificateDescription_status :: Lens' CACertificateDescription (Maybe CACertificateStatus)

The status of a CA certificate.

cACertificateDescription_ownedBy :: Lens' CACertificateDescription (Maybe Text)

The owner of the CA certificate.

cACertificateDescription_lastModifiedDate :: Lens' CACertificateDescription (Maybe UTCTime)

The date the CA certificate was last modified.

cACertificateDescription_certificatePem :: Lens' CACertificateDescription (Maybe Text)

The CA certificate data, in PEM format.

cACertificateDescription_certificateArn :: Lens' CACertificateDescription (Maybe Text)

The CA certificate ARN.

cACertificateDescription_certificateId :: Lens' CACertificateDescription (Maybe Text)

The CA certificate ID.

cACertificateDescription_validity :: Lens' CACertificateDescription (Maybe CertificateValidity)

When the CA certificate is valid.

cACertificateDescription_autoRegistrationStatus :: Lens' CACertificateDescription (Maybe AutoRegistrationStatus)

Whether the CA certificate configured for auto registration of device certificates. Valid values are "ENABLE" and "DISABLE"

cACertificateDescription_creationDate :: Lens' CACertificateDescription (Maybe UTCTime)

The date the CA certificate was created.

cACertificateDescription_generationId :: Lens' CACertificateDescription (Maybe Text)

The generation ID of the CA certificate.

cACertificateDescription_customerVersion :: Lens' CACertificateDescription (Maybe Natural)

The customer version of the CA certificate.