Network.AWS.IoT.Types.Certificate

Description

Documentation

data Certificate

Information about a certificate.

See: newCertificate smart constructor.

Constructors

Certificate'

Fields

status :: Maybe CertificateStatus

The status of the certificate.

The status value REGISTER_INACTIVE is deprecated and should not be used.

certificateArn :: Maybe Text

The ARN of the certificate.

certificateId :: Maybe Text

The ID of the certificate. (The last part of the certificate ARN contains the certificate ID.)

certificateMode :: Maybe CertificateMode

The mode of the certificate.

creationDate :: Maybe POSIX

The date and time the certificate was created.

newCertificate :: Certificate

Create a value of Certificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Certificate', certificate_status - The status of the certificate.

The status value REGISTER_INACTIVE is deprecated and should not be used.

$sel:certificateArn:Certificate', certificate_certificateArn - The ARN of the certificate.

$sel:certificateId:Certificate', certificate_certificateId - The ID of the certificate. (The last part of the certificate ARN contains the certificate ID.)

$sel:certificateMode:Certificate', certificate_certificateMode - The mode of the certificate.

$sel:creationDate:Certificate', certificate_creationDate - The date and time the certificate was created.

certificate_status :: Lens' Certificate (Maybe CertificateStatus)

The status of the certificate.

The status value REGISTER_INACTIVE is deprecated and should not be used.

certificate_certificateArn :: Lens' Certificate (Maybe Text)

The ARN of the certificate.

certificate_certificateId :: Lens' Certificate (Maybe Text)

The ID of the certificate. (The last part of the certificate ARN contains the certificate ID.)

certificate_certificateMode :: Lens' Certificate (Maybe CertificateMode)

The mode of the certificate.

certificate_creationDate :: Lens' Certificate (Maybe UTCTime)

The date and time the certificate was created.