Network.AWS.IoT.Types.CertificateDescription

Description

Documentation

data CertificateDescription

Describes a certificate.

See: newCertificateDescription smart constructor.

Constructors

CertificateDescription'

Fields

lastModifiedDate :: Maybe POSIX

The date and time the certificate was last modified.

status :: Maybe CertificateStatus

The status of the certificate.

certificateMode :: Maybe CertificateMode

The mode of the certificate.

certificateArn :: Maybe Text

The ARN of the certificate.

creationDate :: Maybe POSIX

The date and time the certificate was created.

previousOwnedBy :: Maybe Text

The ID of the Amazon Web Services account of the previous owner of the certificate.

ownedBy :: Maybe Text

The ID of the Amazon Web Services account that owns the certificate.

customerVersion :: Maybe Natural

The customer version of the certificate.

generationId :: Maybe Text

The generation ID of the certificate.

transferData :: Maybe TransferData

The transfer data.

certificateId :: Maybe Text

The ID of the certificate.

certificatePem :: Maybe Text

The certificate data, in PEM format.

caCertificateId :: Maybe Text

The certificate ID of the CA certificate used to sign this certificate.

validity :: Maybe CertificateValidity

When the certificate is valid.

newCertificateDescription :: CertificateDescription

Create a value of CertificateDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:CertificateDescription', certificateDescription_lastModifiedDate - The date and time the certificate was last modified.

$sel:status:CertificateDescription', certificateDescription_status - The status of the certificate.

$sel:certificateMode:CertificateDescription', certificateDescription_certificateMode - The mode of the certificate.

$sel:certificateArn:CertificateDescription', certificateDescription_certificateArn - The ARN of the certificate.

$sel:creationDate:CertificateDescription', certificateDescription_creationDate - The date and time the certificate was created.

$sel:previousOwnedBy:CertificateDescription', certificateDescription_previousOwnedBy - The ID of the Amazon Web Services account of the previous owner of the certificate.

$sel:ownedBy:CertificateDescription', certificateDescription_ownedBy - The ID of the Amazon Web Services account that owns the certificate.

$sel:customerVersion:CertificateDescription', certificateDescription_customerVersion - The customer version of the certificate.

$sel:generationId:CertificateDescription', certificateDescription_generationId - The generation ID of the certificate.

$sel:transferData:CertificateDescription', certificateDescription_transferData - The transfer data.

$sel:certificateId:CertificateDescription', certificateDescription_certificateId - The ID of the certificate.

$sel:certificatePem:CertificateDescription', certificateDescription_certificatePem - The certificate data, in PEM format.

$sel:caCertificateId:CertificateDescription', certificateDescription_caCertificateId - The certificate ID of the CA certificate used to sign this certificate.

$sel:validity:CertificateDescription', certificateDescription_validity - When the certificate is valid.

certificateDescription_lastModifiedDate :: Lens' CertificateDescription (Maybe UTCTime)

The date and time the certificate was last modified.

certificateDescription_status :: Lens' CertificateDescription (Maybe CertificateStatus)

The status of the certificate.

certificateDescription_certificateMode :: Lens' CertificateDescription (Maybe CertificateMode)

The mode of the certificate.

certificateDescription_certificateArn :: Lens' CertificateDescription (Maybe Text)

The ARN of the certificate.

certificateDescription_creationDate :: Lens' CertificateDescription (Maybe UTCTime)

The date and time the certificate was created.

certificateDescription_previousOwnedBy :: Lens' CertificateDescription (Maybe Text)

The ID of the Amazon Web Services account of the previous owner of the certificate.

certificateDescription_ownedBy :: Lens' CertificateDescription (Maybe Text)

The ID of the Amazon Web Services account that owns the certificate.

certificateDescription_customerVersion :: Lens' CertificateDescription (Maybe Natural)

The customer version of the certificate.

certificateDescription_generationId :: Lens' CertificateDescription (Maybe Text)

The generation ID of the certificate.

certificateDescription_transferData :: Lens' CertificateDescription (Maybe TransferData)

The transfer data.

certificateDescription_certificateId :: Lens' CertificateDescription (Maybe Text)

The ID of the certificate.

certificateDescription_certificatePem :: Lens' CertificateDescription (Maybe Text)

The certificate data, in PEM format.

certificateDescription_caCertificateId :: Lens' CertificateDescription (Maybe Text)

The certificate ID of the CA certificate used to sign this certificate.

certificateDescription_validity :: Lens' CertificateDescription (Maybe CertificateValidity)

When the certificate is valid.