Network.AWS.IoT.Types.CustomCodeSigning

Description

Documentation

data CustomCodeSigning

Describes a custom method used to code sign a file.

See: newCustomCodeSigning smart constructor.

Constructors

CustomCodeSigning'

Fields

signature :: Maybe CodeSigningSignature

The signature for the file.

signatureAlgorithm :: Maybe Text

The signature algorithm used to code sign the file.

certificateChain :: Maybe CodeSigningCertificateChain

The certificate chain.

hashAlgorithm :: Maybe Text

The hash algorithm used to code sign the file.

newCustomCodeSigning :: CustomCodeSigning

Create a value of CustomCodeSigning with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:signature:CustomCodeSigning', customCodeSigning_signature - The signature for the file.

$sel:signatureAlgorithm:CustomCodeSigning', customCodeSigning_signatureAlgorithm - The signature algorithm used to code sign the file.

$sel:certificateChain:CustomCodeSigning', customCodeSigning_certificateChain - The certificate chain.

$sel:hashAlgorithm:CustomCodeSigning', customCodeSigning_hashAlgorithm - The hash algorithm used to code sign the file.

customCodeSigning_signature :: Lens' CustomCodeSigning (Maybe CodeSigningSignature)

The signature for the file.

customCodeSigning_signatureAlgorithm :: Lens' CustomCodeSigning (Maybe Text)

The signature algorithm used to code sign the file.

customCodeSigning_certificateChain :: Lens' CustomCodeSigning (Maybe CodeSigningCertificateChain)

The certificate chain.

customCodeSigning_hashAlgorithm :: Lens' CustomCodeSigning (Maybe Text)

The hash algorithm used to code sign the file.