Network.AWS.IoT.Types.DetectMitigationActionExecution

Description

Documentation

data DetectMitigationActionExecution

Describes which mitigation actions should be executed.

See: newDetectMitigationActionExecution smart constructor.

Constructors

DetectMitigationActionExecution'

Fields

status :: Maybe DetectMitigationActionExecutionStatus

The status of a mitigation action.

violationId :: Maybe Text

The unique identifier of the violation.

actionName :: Maybe Text

The friendly name that uniquely identifies the mitigation action.

executionStartDate :: Maybe POSIX

The date a mitigation action was started.

message :: Maybe Text

The message of a mitigation action.

thingName :: Maybe Text

The name of the thing.

taskId :: Maybe Text

The unique identifier of the task.

executionEndDate :: Maybe POSIX

The date a mitigation action ended.

errorCode :: Maybe Text

The error code of a mitigation action.

newDetectMitigationActionExecution :: DetectMitigationActionExecution

Create a value of DetectMitigationActionExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DetectMitigationActionExecution', detectMitigationActionExecution_status - The status of a mitigation action.

$sel:violationId:DetectMitigationActionExecution', detectMitigationActionExecution_violationId - The unique identifier of the violation.

$sel:actionName:DetectMitigationActionExecution', detectMitigationActionExecution_actionName - The friendly name that uniquely identifies the mitigation action.

$sel:executionStartDate:DetectMitigationActionExecution', detectMitigationActionExecution_executionStartDate - The date a mitigation action was started.

$sel:message:DetectMitigationActionExecution', detectMitigationActionExecution_message - The message of a mitigation action.

$sel:thingName:DetectMitigationActionExecution', detectMitigationActionExecution_thingName - The name of the thing.

$sel:taskId:DetectMitigationActionExecution', detectMitigationActionExecution_taskId - The unique identifier of the task.

$sel:executionEndDate:DetectMitigationActionExecution', detectMitigationActionExecution_executionEndDate - The date a mitigation action ended.

$sel:errorCode:DetectMitigationActionExecution', detectMitigationActionExecution_errorCode - The error code of a mitigation action.

detectMitigationActionExecution_status :: Lens' DetectMitigationActionExecution (Maybe DetectMitigationActionExecutionStatus)

The status of a mitigation action.

detectMitigationActionExecution_violationId :: Lens' DetectMitigationActionExecution (Maybe Text)

The unique identifier of the violation.

detectMitigationActionExecution_actionName :: Lens' DetectMitigationActionExecution (Maybe Text)

The friendly name that uniquely identifies the mitigation action.

detectMitigationActionExecution_executionStartDate :: Lens' DetectMitigationActionExecution (Maybe UTCTime)

The date a mitigation action was started.

detectMitigationActionExecution_message :: Lens' DetectMitigationActionExecution (Maybe Text)

The message of a mitigation action.

detectMitigationActionExecution_thingName :: Lens' DetectMitigationActionExecution (Maybe Text)

The name of the thing.

detectMitigationActionExecution_taskId :: Lens' DetectMitigationActionExecution (Maybe Text)

The unique identifier of the task.

detectMitigationActionExecution_executionEndDate :: Lens' DetectMitigationActionExecution (Maybe UTCTime)

The date a mitigation action ended.

detectMitigationActionExecution_errorCode :: Lens' DetectMitigationActionExecution (Maybe Text)

The error code of a mitigation action.