Network.AWS.IoT.Types.DynamoDBAction

Description

Documentation

data DynamoDBAction

Describes an action to write to a DynamoDB table.

The tableName, hashKeyField, and rangeKeyField values must match the values used when you created the table.

The hashKeyValue and rangeKeyvalue fields use a substitution template syntax. These templates provide data at runtime. The syntax is as follows: ${sql-expression}.

You can specify any valid expression in a WHERE or SELECT clause, including JSON properties, comparisons, calculations, and functions. For example, the following field uses the third level of the topic:

"hashKeyValue": "${topic(3)}"

The following field uses the timestamp:

"rangeKeyValue": "${timestamp()}"

See: newDynamoDBAction smart constructor.

Constructors

DynamoDBAction'

Fields

rangeKeyValue :: Maybe Text

The range key value.

rangeKeyType :: Maybe DynamoKeyType

The range key type. Valid values are "STRING" or "NUMBER"

hashKeyType :: Maybe DynamoKeyType

The hash key type. Valid values are "STRING" or "NUMBER"

operation :: Maybe Text

The type of operation to be performed. This follows the substitution template, so it can be ${operation}, but the substitution must result in one of the following: INSERT, UPDATE, or DELETE.

rangeKeyField :: Maybe Text

The range key name.

payloadField :: Maybe Text

The action payload. This name can be customized.

tableName :: Text

The name of the DynamoDB table.

roleArn :: Text

The ARN of the IAM role that grants access to the DynamoDB table.

hashKeyField :: Text

The hash key name.

hashKeyValue :: Text

The hash key value.

newDynamoDBAction

Arguments

:: Text
$sel:tableName:DynamoDBAction'

-> Text
$sel:roleArn:DynamoDBAction'

-> Text
$sel:hashKeyField:DynamoDBAction'

-> Text
$sel:hashKeyValue:DynamoDBAction'

-> DynamoDBAction

Create a value of DynamoDBAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rangeKeyValue:DynamoDBAction', dynamoDBAction_rangeKeyValue - The range key value.

$sel:rangeKeyType:DynamoDBAction', dynamoDBAction_rangeKeyType - The range key type. Valid values are "STRING" or "NUMBER"

$sel:hashKeyType:DynamoDBAction', dynamoDBAction_hashKeyType - The hash key type. Valid values are "STRING" or "NUMBER"

$sel:operation:DynamoDBAction', dynamoDBAction_operation - The type of operation to be performed. This follows the substitution template, so it can be ${operation}, but the substitution must result in one of the following: INSERT, UPDATE, or DELETE.

$sel:rangeKeyField:DynamoDBAction', dynamoDBAction_rangeKeyField - The range key name.

$sel:payloadField:DynamoDBAction', dynamoDBAction_payloadField - The action payload. This name can be customized.

$sel:tableName:DynamoDBAction', dynamoDBAction_tableName - The name of the DynamoDB table.

$sel:roleArn:DynamoDBAction', dynamoDBAction_roleArn - The ARN of the IAM role that grants access to the DynamoDB table.

$sel:hashKeyField:DynamoDBAction', dynamoDBAction_hashKeyField - The hash key name.

$sel:hashKeyValue:DynamoDBAction', dynamoDBAction_hashKeyValue - The hash key value.

dynamoDBAction_rangeKeyValue :: Lens' DynamoDBAction (Maybe Text)

The range key value.

dynamoDBAction_rangeKeyType :: Lens' DynamoDBAction (Maybe DynamoKeyType)

The range key type. Valid values are "STRING" or "NUMBER"

dynamoDBAction_hashKeyType :: Lens' DynamoDBAction (Maybe DynamoKeyType)

The hash key type. Valid values are "STRING" or "NUMBER"

dynamoDBAction_operation :: Lens' DynamoDBAction (Maybe Text)

The type of operation to be performed. This follows the substitution template, so it can be ${operation}, but the substitution must result in one of the following: INSERT, UPDATE, or DELETE.

dynamoDBAction_rangeKeyField :: Lens' DynamoDBAction (Maybe Text)

The range key name.

dynamoDBAction_payloadField :: Lens' DynamoDBAction (Maybe Text)

The action payload. This name can be customized.

dynamoDBAction_tableName :: Lens' DynamoDBAction Text

The name of the DynamoDB table.

dynamoDBAction_roleArn :: Lens' DynamoDBAction Text

The ARN of the IAM role that grants access to the DynamoDB table.

dynamoDBAction_hashKeyField :: Lens' DynamoDBAction Text

The hash key name.

dynamoDBAction_hashKeyValue :: Lens' DynamoDBAction Text

The hash key value.