Network.AWS.IoT.Types.HttpAction

Description

Documentation

data HttpAction

Send data to an HTTPS endpoint.

See: newHttpAction smart constructor.

Constructors

HttpAction'

Fields

headers :: Maybe [HttpActionHeader]

The HTTP headers to send with the message data.

auth :: Maybe HttpAuthorization

The authentication method to use when sending data to an HTTPS endpoint.

confirmationUrl :: Maybe Text

The URL to which IoT sends a confirmation message. The value of the confirmation URL must be a prefix of the endpoint URL. If you do not specify a confirmation URL IoT uses the endpoint URL as the confirmation URL. If you use substitution templates in the confirmationUrl, you must create and enable topic rule destinations that match each possible value of the substitution template before traffic is allowed to your endpoint URL.

url :: Text

The endpoint URL. If substitution templates are used in the URL, you must also specify a confirmationUrl. If this is a new destination, a new TopicRuleDestination is created if possible.

newHttpAction

Arguments

:: Text
$sel:url:HttpAction'

-> HttpAction

Create a value of HttpAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:headers:HttpAction', httpAction_headers - The HTTP headers to send with the message data.

$sel:auth:HttpAction', httpAction_auth - The authentication method to use when sending data to an HTTPS endpoint.

$sel:confirmationUrl:HttpAction', httpAction_confirmationUrl - The URL to which IoT sends a confirmation message. The value of the confirmation URL must be a prefix of the endpoint URL. If you do not specify a confirmation URL IoT uses the endpoint URL as the confirmation URL. If you use substitution templates in the confirmationUrl, you must create and enable topic rule destinations that match each possible value of the substitution template before traffic is allowed to your endpoint URL.

$sel:url:HttpAction', httpAction_url - The endpoint URL. If substitution templates are used in the URL, you must also specify a confirmationUrl. If this is a new destination, a new TopicRuleDestination is created if possible.

httpAction_headers :: Lens' HttpAction (Maybe [HttpActionHeader])

The HTTP headers to send with the message data.

httpAction_auth :: Lens' HttpAction (Maybe HttpAuthorization)

The authentication method to use when sending data to an HTTPS endpoint.

httpAction_confirmationUrl :: Lens' HttpAction (Maybe Text)

The URL to which IoT sends a confirmation message. The value of the confirmation URL must be a prefix of the endpoint URL. If you do not specify a confirmation URL IoT uses the endpoint URL as the confirmation URL. If you use substitution templates in the confirmationUrl, you must create and enable topic rule destinations that match each possible value of the substitution template before traffic is allowed to your endpoint URL.

httpAction_url :: Lens' HttpAction Text

The endpoint URL. If substitution templates are used in the URL, you must also specify a confirmationUrl. If this is a new destination, a new TopicRuleDestination is created if possible.