Network.AWS.IoT.Types.IotAnalyticsAction

Description

Documentation

data IotAnalyticsAction

Sends message data to an IoT Analytics channel.

See: newIotAnalyticsAction smart constructor.

Constructors

IotAnalyticsAction'

Fields

channelName :: Maybe Text

The name of the IoT Analytics channel to which message data will be sent.

roleArn :: Maybe Text

The ARN of the role which has a policy that grants IoT Analytics permission to send message data via IoT Analytics (iotanalytics:BatchPutMessage).

batchMode :: Maybe Bool

Whether to process the action as a batch. The default value is false.

When batchMode is true and the rule SQL statement evaluates to an Array, each Array element is delivered as a separate message when passed by BatchPutMessage to the IoT Analytics channel. The resulting array can't have more than 100 messages.

channelArn :: Maybe Text

(deprecated) The ARN of the IoT Analytics channel to which message data will be sent.

newIotAnalyticsAction :: IotAnalyticsAction

Create a value of IotAnalyticsAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelName:IotAnalyticsAction', iotAnalyticsAction_channelName - The name of the IoT Analytics channel to which message data will be sent.

$sel:roleArn:IotAnalyticsAction', iotAnalyticsAction_roleArn - The ARN of the role which has a policy that grants IoT Analytics permission to send message data via IoT Analytics (iotanalytics:BatchPutMessage).

$sel:batchMode:IotAnalyticsAction', iotAnalyticsAction_batchMode - Whether to process the action as a batch. The default value is false.

When batchMode is true and the rule SQL statement evaluates to an Array, each Array element is delivered as a separate message when passed by BatchPutMessage to the IoT Analytics channel. The resulting array can't have more than 100 messages.

$sel:channelArn:IotAnalyticsAction', iotAnalyticsAction_channelArn - (deprecated) The ARN of the IoT Analytics channel to which message data will be sent.

iotAnalyticsAction_channelName :: Lens' IotAnalyticsAction (Maybe Text)

The name of the IoT Analytics channel to which message data will be sent.

iotAnalyticsAction_roleArn :: Lens' IotAnalyticsAction (Maybe Text)

The ARN of the role which has a policy that grants IoT Analytics permission to send message data via IoT Analytics (iotanalytics:BatchPutMessage).

iotAnalyticsAction_batchMode :: Lens' IotAnalyticsAction (Maybe Bool)

Whether to process the action as a batch. The default value is false.

When batchMode is true and the rule SQL statement evaluates to an Array, each Array element is delivered as a separate message when passed by BatchPutMessage to the IoT Analytics channel. The resulting array can't have more than 100 messages.

iotAnalyticsAction_channelArn :: Lens' IotAnalyticsAction (Maybe Text)

(deprecated) The ARN of the IoT Analytics channel to which message data will be sent.