Network.AWS.IoT.Types.IotEventsAction

Description

Documentation

data IotEventsAction

Sends an input to an IoT Events detector.

See: newIotEventsAction smart constructor.

Constructors

IotEventsAction'

Fields

batchMode :: Maybe Bool

Whether to process the event actions as a batch. The default value is false.

When batchMode is true, you can't specify a messageId.

When batchMode is true and the rule SQL statement evaluates to an Array, each Array element is treated as a separate message when it's sent to IoT Events by calling BatchPutMessage . The resulting array can't have more than 10 messages.

messageId :: Maybe Text

The ID of the message. The default messageId is a new UUID value.

When batchMode is true, you can't specify a messageId--a new UUID value will be assigned.

Assign a value to this property to ensure that only one input (message) with a given messageId will be processed by an IoT Events detector.

inputName :: Text

The name of the IoT Events input.

roleArn :: Text

The ARN of the role that grants IoT permission to send an input to an IoT Events detector. ("Action":"iotevents:BatchPutMessage").

newIotEventsAction

Arguments

:: Text
$sel:inputName:IotEventsAction'

-> Text
$sel:roleArn:IotEventsAction'

-> IotEventsAction

Create a value of IotEventsAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:batchMode:IotEventsAction', iotEventsAction_batchMode - Whether to process the event actions as a batch. The default value is false.

When batchMode is true, you can't specify a messageId.

When batchMode is true and the rule SQL statement evaluates to an Array, each Array element is treated as a separate message when it's sent to IoT Events by calling BatchPutMessage . The resulting array can't have more than 10 messages.

$sel:messageId:IotEventsAction', iotEventsAction_messageId - The ID of the message. The default messageId is a new UUID value.

When batchMode is true, you can't specify a messageId--a new UUID value will be assigned.

Assign a value to this property to ensure that only one input (message) with a given messageId will be processed by an IoT Events detector.

$sel:inputName:IotEventsAction', iotEventsAction_inputName - The name of the IoT Events input.

$sel:roleArn:IotEventsAction', iotEventsAction_roleArn - The ARN of the role that grants IoT permission to send an input to an IoT Events detector. ("Action":"iotevents:BatchPutMessage").

iotEventsAction_batchMode :: Lens' IotEventsAction (Maybe Bool)

Whether to process the event actions as a batch. The default value is false.

When batchMode is true, you can't specify a messageId.

When batchMode is true and the rule SQL statement evaluates to an Array, each Array element is treated as a separate message when it's sent to IoT Events by calling BatchPutMessage . The resulting array can't have more than 10 messages.

iotEventsAction_messageId :: Lens' IotEventsAction (Maybe Text)

The ID of the message. The default messageId is a new UUID value.

When batchMode is true, you can't specify a messageId--a new UUID value will be assigned.

Assign a value to this property to ensure that only one input (message) with a given messageId will be processed by an IoT Events detector.

iotEventsAction_inputName :: Lens' IotEventsAction Text

The name of the IoT Events input.

iotEventsAction_roleArn :: Lens' IotEventsAction Text

The ARN of the role that grants IoT permission to send an input to an IoT Events detector. ("Action":"iotevents:BatchPutMessage").