Network.AWS.IoT.Types.Job

Description

Documentation

data Job

The Job object contains details about a job.

See: newJob smart constructor.

Constructors

Job'

Fields

jobExecutionsRolloutConfig :: Maybe JobExecutionsRolloutConfig

Allows you to create a staged rollout of a job.

status :: Maybe JobStatus

The status of the job, one of IN_PROGRESS, CANCELED, DELETION_IN_PROGRESS or COMPLETED.

reasonCode :: Maybe Text

If the job was updated, provides the reason code for the update.

timeoutConfig :: Maybe TimeoutConfig

Specifies the amount of time each device has to finish its execution of the job. A timer is started when the job execution status is set to IN_PROGRESS. If the job execution status is not set to another terminal state before the timer expires, it will be automatically set to TIMED_OUT.

targetSelection :: Maybe TargetSelection

Specifies whether the job will continue to run (CONTINUOUS), or will be complete after all those things specified as targets have completed the job (SNAPSHOT). If continuous, the job may also be run on a thing when a change is detected in a target. For example, a job will run on a device when the thing representing the device is added to a target group, even after the job was completed by all things originally in the group.

namespaceId :: Maybe Text

The namespace used to indicate that a job is a customer-managed job.

When you specify a value for this parameter, Amazon Web Services IoT Core sends jobs notifications to MQTT topics that contain the value in the following format.

$aws/things/THING_NAME/jobs/JOB_ID/notify-namespace-NAMESPACE_ID/

The namespaceId feature is in public preview.

jobProcessDetails :: Maybe JobProcessDetails

Details about the job process.

comment :: Maybe Text

If the job was updated, describes the reason for the update.

completedAt :: Maybe POSIX

The time, in seconds since the epoch, when the job was completed.

createdAt :: Maybe POSIX

The time, in seconds since the epoch, when the job was created.

forceCanceled :: Maybe Bool

Will be true if the job was canceled with the optional force parameter set to true.

jobArn :: Maybe Text

An ARN identifying the job with format "arn:aws:iot:region:account:job/jobId".

targets :: Maybe (NonEmpty Text)

A list of IoT things and thing groups to which the job should be sent.

presignedUrlConfig :: Maybe PresignedUrlConfig

Configuration for pre-signed S3 URLs.

description :: Maybe Text

A short text description of the job.

abortConfig :: Maybe AbortConfig

Configuration for criteria to abort the job.

jobTemplateArn :: Maybe Text

The ARN of the job template used to create the job.

jobId :: Maybe Text

The unique identifier you assigned to this job when it was created.

lastUpdatedAt :: Maybe POSIX

The time, in seconds since the epoch, when the job was last updated.

newJob :: Job

Create a value of Job with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobExecutionsRolloutConfig:Job', job_jobExecutionsRolloutConfig - Allows you to create a staged rollout of a job.

$sel:status:Job', job_status - The status of the job, one of IN_PROGRESS, CANCELED, DELETION_IN_PROGRESS or COMPLETED.

$sel:reasonCode:Job', job_reasonCode - If the job was updated, provides the reason code for the update.

$sel:timeoutConfig:Job', job_timeoutConfig - Specifies the amount of time each device has to finish its execution of the job. A timer is started when the job execution status is set to IN_PROGRESS. If the job execution status is not set to another terminal state before the timer expires, it will be automatically set to TIMED_OUT.

$sel:targetSelection:Job', job_targetSelection - Specifies whether the job will continue to run (CONTINUOUS), or will be complete after all those things specified as targets have completed the job (SNAPSHOT). If continuous, the job may also be run on a thing when a change is detected in a target. For example, a job will run on a device when the thing representing the device is added to a target group, even after the job was completed by all things originally in the group.

$sel:namespaceId:Job', job_namespaceId - The namespace used to indicate that a job is a customer-managed job.

When you specify a value for this parameter, Amazon Web Services IoT Core sends jobs notifications to MQTT topics that contain the value in the following format.

$aws/things/THING_NAME/jobs/JOB_ID/notify-namespace-NAMESPACE_ID/

The namespaceId feature is in public preview.

$sel:jobProcessDetails:Job', job_jobProcessDetails - Details about the job process.

$sel:comment:Job', job_comment - If the job was updated, describes the reason for the update.

$sel:completedAt:Job', job_completedAt - The time, in seconds since the epoch, when the job was completed.

$sel:createdAt:Job', job_createdAt - The time, in seconds since the epoch, when the job was created.

$sel:forceCanceled:Job', job_forceCanceled - Will be true if the job was canceled with the optional force parameter set to true.

$sel:jobArn:Job', job_jobArn - An ARN identifying the job with format "arn:aws:iot:region:account:job/jobId".

$sel:targets:Job', job_targets - A list of IoT things and thing groups to which the job should be sent.

$sel:presignedUrlConfig:Job', job_presignedUrlConfig - Configuration for pre-signed S3 URLs.

$sel:description:Job', job_description - A short text description of the job.

$sel:abortConfig:Job', job_abortConfig - Configuration for criteria to abort the job.

$sel:jobTemplateArn:Job', job_jobTemplateArn - The ARN of the job template used to create the job.

$sel:jobId:Job', job_jobId - The unique identifier you assigned to this job when it was created.

$sel:lastUpdatedAt:Job', job_lastUpdatedAt - The time, in seconds since the epoch, when the job was last updated.

job_jobExecutionsRolloutConfig :: Lens' Job (Maybe JobExecutionsRolloutConfig)

Allows you to create a staged rollout of a job.

job_status :: Lens' Job (Maybe JobStatus)

The status of the job, one of IN_PROGRESS, CANCELED, DELETION_IN_PROGRESS or COMPLETED.

job_reasonCode :: Lens' Job (Maybe Text)

If the job was updated, provides the reason code for the update.

job_timeoutConfig :: Lens' Job (Maybe TimeoutConfig)

Specifies the amount of time each device has to finish its execution of the job. A timer is started when the job execution status is set to IN_PROGRESS. If the job execution status is not set to another terminal state before the timer expires, it will be automatically set to TIMED_OUT.

job_targetSelection :: Lens' Job (Maybe TargetSelection)

Specifies whether the job will continue to run (CONTINUOUS), or will be complete after all those things specified as targets have completed the job (SNAPSHOT). If continuous, the job may also be run on a thing when a change is detected in a target. For example, a job will run on a device when the thing representing the device is added to a target group, even after the job was completed by all things originally in the group.

job_namespaceId :: Lens' Job (Maybe Text)

The namespace used to indicate that a job is a customer-managed job.

When you specify a value for this parameter, Amazon Web Services IoT Core sends jobs notifications to MQTT topics that contain the value in the following format.

$aws/things/THING_NAME/jobs/JOB_ID/notify-namespace-NAMESPACE_ID/

The namespaceId feature is in public preview.

job_jobProcessDetails :: Lens' Job (Maybe JobProcessDetails)

Details about the job process.

job_comment :: Lens' Job (Maybe Text)

If the job was updated, describes the reason for the update.

job_completedAt :: Lens' Job (Maybe UTCTime)

The time, in seconds since the epoch, when the job was completed.

job_createdAt :: Lens' Job (Maybe UTCTime)

The time, in seconds since the epoch, when the job was created.

job_forceCanceled :: Lens' Job (Maybe Bool)

Will be true if the job was canceled with the optional force parameter set to true.

job_jobArn :: Lens' Job (Maybe Text)

An ARN identifying the job with format "arn:aws:iot:region:account:job/jobId".

job_targets :: Lens' Job (Maybe (NonEmpty Text))

A list of IoT things and thing groups to which the job should be sent.

job_presignedUrlConfig :: Lens' Job (Maybe PresignedUrlConfig)

Configuration for pre-signed S3 URLs.

job_description :: Lens' Job (Maybe Text)

A short text description of the job.

job_abortConfig :: Lens' Job (Maybe AbortConfig)

Configuration for criteria to abort the job.

job_jobTemplateArn :: Lens' Job (Maybe Text)

The ARN of the job template used to create the job.

job_jobId :: Lens' Job (Maybe Text)

The unique identifier you assigned to this job when it was created.

job_lastUpdatedAt :: Lens' Job (Maybe UTCTime)

The time, in seconds since the epoch, when the job was last updated.