Network.AWS.IoT.Types.JobExecution

Description

Documentation

data JobExecution

The job execution object represents the execution of a job on a particular device.

See: newJobExecution smart constructor.

Constructors

JobExecution'

Fields

startedAt :: Maybe POSIX

The time, in seconds since the epoch, when the job execution started.

status :: Maybe JobExecutionStatus

The status of the job execution (IN_PROGRESS, QUEUED, FAILED, SUCCEEDED, TIMED_OUT, CANCELED, or REJECTED).

thingArn :: Maybe Text

The ARN of the thing on which the job execution is running.

statusDetails :: Maybe JobExecutionStatusDetails

A collection of name/value pairs that describe the status of the job execution.

queuedAt :: Maybe POSIX

The time, in seconds since the epoch, when the job execution was queued.

forceCanceled :: Maybe Bool

Will be true if the job execution was canceled with the optional force parameter set to true.

executionNumber :: Maybe Integer

A string (consisting of the digits "0" through "9") which identifies this particular job execution on this particular device. It can be used in commands which return or update job execution information.

versionNumber :: Maybe Integer

The version of the job execution. Job execution versions are incremented each time they are updated by a device.

approximateSecondsBeforeTimedOut :: Maybe Integer

The estimated number of seconds that remain before the job execution status will be changed to TIMED_OUT. The timeout interval can be anywhere between 1 minute and 7 days (1 to 10080 minutes). The actual job execution timeout can occur up to 60 seconds later than the estimated duration. This value will not be included if the job execution has reached a terminal status.

jobId :: Maybe Text

The unique identifier you assigned to the job when it was created.

lastUpdatedAt :: Maybe POSIX

The time, in seconds since the epoch, when the job execution was last updated.

newJobExecution :: JobExecution

Create a value of JobExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startedAt:JobExecution', jobExecution_startedAt - The time, in seconds since the epoch, when the job execution started.

$sel:status:JobExecution', jobExecution_status - The status of the job execution (IN_PROGRESS, QUEUED, FAILED, SUCCEEDED, TIMED_OUT, CANCELED, or REJECTED).

$sel:thingArn:JobExecution', jobExecution_thingArn - The ARN of the thing on which the job execution is running.

$sel:statusDetails:JobExecution', jobExecution_statusDetails - A collection of name/value pairs that describe the status of the job execution.

$sel:queuedAt:JobExecution', jobExecution_queuedAt - The time, in seconds since the epoch, when the job execution was queued.

$sel:forceCanceled:JobExecution', jobExecution_forceCanceled - Will be true if the job execution was canceled with the optional force parameter set to true.

$sel:executionNumber:JobExecution', jobExecution_executionNumber - A string (consisting of the digits "0" through "9") which identifies this particular job execution on this particular device. It can be used in commands which return or update job execution information.

$sel:versionNumber:JobExecution', jobExecution_versionNumber - The version of the job execution. Job execution versions are incremented each time they are updated by a device.

$sel:approximateSecondsBeforeTimedOut:JobExecution', jobExecution_approximateSecondsBeforeTimedOut - The estimated number of seconds that remain before the job execution status will be changed to TIMED_OUT. The timeout interval can be anywhere between 1 minute and 7 days (1 to 10080 minutes). The actual job execution timeout can occur up to 60 seconds later than the estimated duration. This value will not be included if the job execution has reached a terminal status.

$sel:jobId:JobExecution', jobExecution_jobId - The unique identifier you assigned to the job when it was created.

$sel:lastUpdatedAt:JobExecution', jobExecution_lastUpdatedAt - The time, in seconds since the epoch, when the job execution was last updated.

jobExecution_startedAt :: Lens' JobExecution (Maybe UTCTime)

The time, in seconds since the epoch, when the job execution started.

jobExecution_status :: Lens' JobExecution (Maybe JobExecutionStatus)

The status of the job execution (IN_PROGRESS, QUEUED, FAILED, SUCCEEDED, TIMED_OUT, CANCELED, or REJECTED).

jobExecution_thingArn :: Lens' JobExecution (Maybe Text)

The ARN of the thing on which the job execution is running.

jobExecution_statusDetails :: Lens' JobExecution (Maybe JobExecutionStatusDetails)

A collection of name/value pairs that describe the status of the job execution.

jobExecution_queuedAt :: Lens' JobExecution (Maybe UTCTime)

The time, in seconds since the epoch, when the job execution was queued.

jobExecution_forceCanceled :: Lens' JobExecution (Maybe Bool)

Will be true if the job execution was canceled with the optional force parameter set to true.

jobExecution_executionNumber :: Lens' JobExecution (Maybe Integer)

A string (consisting of the digits "0" through "9") which identifies this particular job execution on this particular device. It can be used in commands which return or update job execution information.

jobExecution_versionNumber :: Lens' JobExecution (Maybe Integer)

The version of the job execution. Job execution versions are incremented each time they are updated by a device.

jobExecution_approximateSecondsBeforeTimedOut :: Lens' JobExecution (Maybe Integer)

The estimated number of seconds that remain before the job execution status will be changed to TIMED_OUT. The timeout interval can be anywhere between 1 minute and 7 days (1 to 10080 minutes). The actual job execution timeout can occur up to 60 seconds later than the estimated duration. This value will not be included if the job execution has reached a terminal status.

jobExecution_jobId :: Lens' JobExecution (Maybe Text)

The unique identifier you assigned to the job when it was created.

jobExecution_lastUpdatedAt :: Lens' JobExecution (Maybe UTCTime)

The time, in seconds since the epoch, when the job execution was last updated.