Network.AWS.IoT.Types.JobProcessDetails

Description

Documentation

data JobProcessDetails

The job process details.

See: newJobProcessDetails smart constructor.

Constructors

JobProcessDetails'

Fields

processingTargets :: Maybe [Text]

The target devices to which the job execution is being rolled out. This value will be null after the job execution has finished rolling out to all the target devices.

numberOfSucceededThings :: Maybe Int

The number of things which successfully completed the job.

numberOfQueuedThings :: Maybe Int

The number of things that are awaiting execution of the job.

numberOfInProgressThings :: Maybe Int

The number of things currently executing the job.

numberOfRemovedThings :: Maybe Int

The number of things that are no longer scheduled to execute the job because they have been deleted or have been removed from the group that was a target of the job.

numberOfFailedThings :: Maybe Int

The number of things that failed executing the job.

numberOfRejectedThings :: Maybe Int

The number of things that rejected the job.

numberOfTimedOutThings :: Maybe Int

The number of things whose job execution status is TIMED_OUT.

numberOfCanceledThings :: Maybe Int

The number of things that cancelled the job.

newJobProcessDetails :: JobProcessDetails

Create a value of JobProcessDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:processingTargets:JobProcessDetails', jobProcessDetails_processingTargets - The target devices to which the job execution is being rolled out. This value will be null after the job execution has finished rolling out to all the target devices.

$sel:numberOfSucceededThings:JobProcessDetails', jobProcessDetails_numberOfSucceededThings - The number of things which successfully completed the job.

$sel:numberOfQueuedThings:JobProcessDetails', jobProcessDetails_numberOfQueuedThings - The number of things that are awaiting execution of the job.

$sel:numberOfInProgressThings:JobProcessDetails', jobProcessDetails_numberOfInProgressThings - The number of things currently executing the job.

$sel:numberOfRemovedThings:JobProcessDetails', jobProcessDetails_numberOfRemovedThings - The number of things that are no longer scheduled to execute the job because they have been deleted or have been removed from the group that was a target of the job.

$sel:numberOfFailedThings:JobProcessDetails', jobProcessDetails_numberOfFailedThings - The number of things that failed executing the job.

$sel:numberOfRejectedThings:JobProcessDetails', jobProcessDetails_numberOfRejectedThings - The number of things that rejected the job.

$sel:numberOfTimedOutThings:JobProcessDetails', jobProcessDetails_numberOfTimedOutThings - The number of things whose job execution status is TIMED_OUT.

$sel:numberOfCanceledThings:JobProcessDetails', jobProcessDetails_numberOfCanceledThings - The number of things that cancelled the job.

jobProcessDetails_processingTargets :: Lens' JobProcessDetails (Maybe [Text])

The target devices to which the job execution is being rolled out. This value will be null after the job execution has finished rolling out to all the target devices.

jobProcessDetails_numberOfSucceededThings :: Lens' JobProcessDetails (Maybe Int)

The number of things which successfully completed the job.

jobProcessDetails_numberOfQueuedThings :: Lens' JobProcessDetails (Maybe Int)

The number of things that are awaiting execution of the job.

jobProcessDetails_numberOfInProgressThings :: Lens' JobProcessDetails (Maybe Int)

The number of things currently executing the job.

jobProcessDetails_numberOfRemovedThings :: Lens' JobProcessDetails (Maybe Int)

The number of things that are no longer scheduled to execute the job because they have been deleted or have been removed from the group that was a target of the job.

jobProcessDetails_numberOfFailedThings :: Lens' JobProcessDetails (Maybe Int)

The number of things that failed executing the job.

jobProcessDetails_numberOfRejectedThings :: Lens' JobProcessDetails (Maybe Int)

The number of things that rejected the job.

jobProcessDetails_numberOfTimedOutThings :: Lens' JobProcessDetails (Maybe Int)

The number of things whose job execution status is TIMED_OUT.

jobProcessDetails_numberOfCanceledThings :: Lens' JobProcessDetails (Maybe Int)

The number of things that cancelled the job.