Network.AWS.IoT.Types.JobSummary

Description

Documentation

data JobSummary

The job summary.

See: newJobSummary smart constructor.

Constructors

JobSummary'

Fields

status :: Maybe JobStatus

The job summary status.

targetSelection :: Maybe TargetSelection

Specifies whether the job will continue to run (CONTINUOUS), or will be complete after all those things specified as targets have completed the job (SNAPSHOT). If continuous, the job may also be run on a thing when a change is detected in a target. For example, a job will run on a thing when the thing is added to a target group, even after the job was completed by all things originally in the group.

completedAt :: Maybe POSIX

The time, in seconds since the epoch, when the job completed.

createdAt :: Maybe POSIX

The time, in seconds since the epoch, when the job was created.

jobArn :: Maybe Text

The job ARN.

thingGroupId :: Maybe Text

The ID of the thing group.

jobId :: Maybe Text

The unique identifier you assigned to this job when it was created.

lastUpdatedAt :: Maybe POSIX

The time, in seconds since the epoch, when the job was last updated.

newJobSummary :: JobSummary

Create a value of JobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:JobSummary', jobSummary_status - The job summary status.

$sel:targetSelection:JobSummary', jobSummary_targetSelection - Specifies whether the job will continue to run (CONTINUOUS), or will be complete after all those things specified as targets have completed the job (SNAPSHOT). If continuous, the job may also be run on a thing when a change is detected in a target. For example, a job will run on a thing when the thing is added to a target group, even after the job was completed by all things originally in the group.

$sel:completedAt:JobSummary', jobSummary_completedAt - The time, in seconds since the epoch, when the job completed.

$sel:createdAt:JobSummary', jobSummary_createdAt - The time, in seconds since the epoch, when the job was created.

$sel:jobArn:JobSummary', jobSummary_jobArn - The job ARN.

$sel:thingGroupId:JobSummary', jobSummary_thingGroupId - The ID of the thing group.

$sel:jobId:JobSummary', jobSummary_jobId - The unique identifier you assigned to this job when it was created.

$sel:lastUpdatedAt:JobSummary', jobSummary_lastUpdatedAt - The time, in seconds since the epoch, when the job was last updated.

jobSummary_status :: Lens' JobSummary (Maybe JobStatus)

The job summary status.

jobSummary_targetSelection :: Lens' JobSummary (Maybe TargetSelection)

Specifies whether the job will continue to run (CONTINUOUS), or will be complete after all those things specified as targets have completed the job (SNAPSHOT). If continuous, the job may also be run on a thing when a change is detected in a target. For example, a job will run on a thing when the thing is added to a target group, even after the job was completed by all things originally in the group.

jobSummary_completedAt :: Lens' JobSummary (Maybe UTCTime)

The time, in seconds since the epoch, when the job completed.

jobSummary_createdAt :: Lens' JobSummary (Maybe UTCTime)

The time, in seconds since the epoch, when the job was created.

jobSummary_jobArn :: Lens' JobSummary (Maybe Text)

The job ARN.

jobSummary_thingGroupId :: Lens' JobSummary (Maybe Text)

The ID of the thing group.

jobSummary_jobId :: Lens' JobSummary (Maybe Text)

The unique identifier you assigned to this job when it was created.

jobSummary_lastUpdatedAt :: Lens' JobSummary (Maybe UTCTime)

The time, in seconds since the epoch, when the job was last updated.