Network.AWS.IoT.Types.MetricValue

Description

Documentation

data MetricValue

The value to be compared with the metric.

See: newMetricValue smart constructor.

Constructors

MetricValue'

Fields

numbers :: Maybe [Double]

The numeral values of a metric.

ports :: Maybe [Natural]

If the comparisonOperator calls for a set of ports, use this to specify that set to be compared with the metric.

cidrs :: Maybe [Text]

If the comparisonOperator calls for a set of CIDRs, use this to specify that set to be compared with the metric.

strings :: Maybe [Text]

The string values of a metric.

number :: Maybe Double

The numeral value of a metric.

count :: Maybe Natural

If the comparisonOperator calls for a numeric value, use this to specify that numeric value to be compared with the metric.

newMetricValue :: MetricValue

Create a value of MetricValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:numbers:MetricValue', metricValue_numbers - The numeral values of a metric.

$sel:ports:MetricValue', metricValue_ports - If the comparisonOperator calls for a set of ports, use this to specify that set to be compared with the metric.

$sel:cidrs:MetricValue', metricValue_cidrs - If the comparisonOperator calls for a set of CIDRs, use this to specify that set to be compared with the metric.

$sel:strings:MetricValue', metricValue_strings - The string values of a metric.

$sel:number:MetricValue', metricValue_number - The numeral value of a metric.

$sel:count:MetricValue', metricValue_count - If the comparisonOperator calls for a numeric value, use this to specify that numeric value to be compared with the metric.

metricValue_numbers :: Lens' MetricValue (Maybe [Double])

The numeral values of a metric.

metricValue_ports :: Lens' MetricValue (Maybe [Natural])

If the comparisonOperator calls for a set of ports, use this to specify that set to be compared with the metric.

metricValue_cidrs :: Lens' MetricValue (Maybe [Text])

If the comparisonOperator calls for a set of CIDRs, use this to specify that set to be compared with the metric.

metricValue_strings :: Lens' MetricValue (Maybe [Text])

The string values of a metric.

metricValue_number :: Lens' MetricValue (Maybe Double)

The numeral value of a metric.

metricValue_count :: Lens' MetricValue (Maybe Natural)

If the comparisonOperator calls for a numeric value, use this to specify that numeric value to be compared with the metric.