Network.AWS.IoT.Types.MitigationActionParams

Description

Documentation

data MitigationActionParams

The set of parameters for this mitigation action. You can specify only one type of parameter (in other words, you can apply only one action for each defined mitigation action).

See: newMitigationActionParams smart constructor.

Constructors

MitigationActionParams'

Fields

enableIoTLoggingParams :: Maybe EnableIoTLoggingParams

Parameters to define a mitigation action that enables Amazon Web Services IoT Core logging at a specified level of detail.

addThingsToThingGroupParams :: Maybe AddThingsToThingGroupParams

Parameters to define a mitigation action that moves devices associated with a certificate to one or more specified thing groups, typically for quarantine.

updateCACertificateParams :: Maybe UpdateCACertificateParams

Parameters to define a mitigation action that changes the state of the CA certificate to inactive.

updateDeviceCertificateParams :: Maybe UpdateDeviceCertificateParams

Parameters to define a mitigation action that changes the state of the device certificate to inactive.

replaceDefaultPolicyVersionParams :: Maybe ReplaceDefaultPolicyVersionParams

Parameters to define a mitigation action that adds a blank policy to restrict permissions.

publishFindingToSnsParams :: Maybe PublishFindingToSnsParams

Parameters to define a mitigation action that publishes findings to Amazon Simple Notification Service (Amazon SNS. You can implement your own custom actions in response to the Amazon SNS messages.

newMitigationActionParams :: MitigationActionParams

Create a value of MitigationActionParams with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enableIoTLoggingParams:MitigationActionParams', mitigationActionParams_enableIoTLoggingParams - Parameters to define a mitigation action that enables Amazon Web Services IoT Core logging at a specified level of detail.

$sel:addThingsToThingGroupParams:MitigationActionParams', mitigationActionParams_addThingsToThingGroupParams - Parameters to define a mitigation action that moves devices associated with a certificate to one or more specified thing groups, typically for quarantine.

$sel:updateCACertificateParams:MitigationActionParams', mitigationActionParams_updateCACertificateParams - Parameters to define a mitigation action that changes the state of the CA certificate to inactive.

$sel:updateDeviceCertificateParams:MitigationActionParams', mitigationActionParams_updateDeviceCertificateParams - Parameters to define a mitigation action that changes the state of the device certificate to inactive.

$sel:replaceDefaultPolicyVersionParams:MitigationActionParams', mitigationActionParams_replaceDefaultPolicyVersionParams - Parameters to define a mitigation action that adds a blank policy to restrict permissions.

$sel:publishFindingToSnsParams:MitigationActionParams', mitigationActionParams_publishFindingToSnsParams - Parameters to define a mitigation action that publishes findings to Amazon Simple Notification Service (Amazon SNS. You can implement your own custom actions in response to the Amazon SNS messages.

mitigationActionParams_enableIoTLoggingParams :: Lens' MitigationActionParams (Maybe EnableIoTLoggingParams)

Parameters to define a mitigation action that enables Amazon Web Services IoT Core logging at a specified level of detail.

mitigationActionParams_addThingsToThingGroupParams :: Lens' MitigationActionParams (Maybe AddThingsToThingGroupParams)

Parameters to define a mitigation action that moves devices associated with a certificate to one or more specified thing groups, typically for quarantine.

mitigationActionParams_updateCACertificateParams :: Lens' MitigationActionParams (Maybe UpdateCACertificateParams)

Parameters to define a mitigation action that changes the state of the CA certificate to inactive.

mitigationActionParams_updateDeviceCertificateParams :: Lens' MitigationActionParams (Maybe UpdateDeviceCertificateParams)

Parameters to define a mitigation action that changes the state of the device certificate to inactive.

mitigationActionParams_replaceDefaultPolicyVersionParams :: Lens' MitigationActionParams (Maybe ReplaceDefaultPolicyVersionParams)

Parameters to define a mitigation action that adds a blank policy to restrict permissions.

mitigationActionParams_publishFindingToSnsParams :: Lens' MitigationActionParams (Maybe PublishFindingToSnsParams)

Parameters to define a mitigation action that publishes findings to Amazon Simple Notification Service (Amazon SNS. You can implement your own custom actions in response to the Amazon SNS messages.