Network.AWS.IoT.Types.OTAUpdateFile

Description

Documentation

data OTAUpdateFile

Describes a file to be associated with an OTA update.

See: newOTAUpdateFile smart constructor.

Constructors

OTAUpdateFile'

Fields

fileLocation :: Maybe FileLocation

The location of the updated firmware.

fileType :: Maybe Natural

An integer value you can include in the job document to allow your devices to identify the type of file received from the cloud.

fileVersion :: Maybe Text

The file version.

attributes :: Maybe (HashMap Text Text)

A list of name/attribute pairs.

codeSigning :: Maybe CodeSigning

The code signing method of the file.

fileName :: Maybe Text

The name of the file.

newOTAUpdateFile :: OTAUpdateFile

Create a value of OTAUpdateFile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fileLocation:OTAUpdateFile', oTAUpdateFile_fileLocation - The location of the updated firmware.

$sel:fileType:OTAUpdateFile', oTAUpdateFile_fileType - An integer value you can include in the job document to allow your devices to identify the type of file received from the cloud.

$sel:fileVersion:OTAUpdateFile', oTAUpdateFile_fileVersion - The file version.

$sel:attributes:OTAUpdateFile', oTAUpdateFile_attributes - A list of name/attribute pairs.

$sel:codeSigning:OTAUpdateFile', oTAUpdateFile_codeSigning - The code signing method of the file.

$sel:fileName:OTAUpdateFile', oTAUpdateFile_fileName - The name of the file.

oTAUpdateFile_fileLocation :: Lens' OTAUpdateFile (Maybe FileLocation)

The location of the updated firmware.

oTAUpdateFile_fileType :: Lens' OTAUpdateFile (Maybe Natural)

An integer value you can include in the job document to allow your devices to identify the type of file received from the cloud.

oTAUpdateFile_fileVersion :: Lens' OTAUpdateFile (Maybe Text)

The file version.

oTAUpdateFile_attributes :: Lens' OTAUpdateFile (Maybe (HashMap Text Text))

A list of name/attribute pairs.

oTAUpdateFile_codeSigning :: Lens' OTAUpdateFile (Maybe CodeSigning)

The code signing method of the file.

oTAUpdateFile_fileName :: Lens' OTAUpdateFile (Maybe Text)

The name of the file.