Network.AWS.IoT.Types.OTAUpdateInfo

Description

Documentation

data OTAUpdateInfo

Information about an OTA update.

See: newOTAUpdateInfo smart constructor.

Constructors

OTAUpdateInfo'

Fields

otaUpdateStatus :: Maybe OTAUpdateStatus

The status of the OTA update.

lastModifiedDate :: Maybe POSIX

The date when the OTA update was last updated.

targetSelection :: Maybe TargetSelection

Specifies whether the OTA update will continue to run (CONTINUOUS), or will be complete after all those things specified as targets have completed the OTA update (SNAPSHOT). If continuous, the OTA update may also be run on a thing when a change is detected in a target. For example, an OTA update will run on a thing when the thing is added to a target group, even after the OTA update was completed by all things originally in the group.

otaUpdateArn :: Maybe Text

The OTA update ARN.

creationDate :: Maybe POSIX

The date when the OTA update was created.

protocols :: Maybe (NonEmpty Protocol)

The protocol used to transfer the OTA update image. Valid values are [HTTP], [MQTT], [HTTP, MQTT]. When both HTTP and MQTT are specified, the target device can choose the protocol.

awsIotJobArn :: Maybe Text

The IoT job ARN associated with the OTA update.

awsIotJobId :: Maybe Text

The IoT job ID associated with the OTA update.

targets :: Maybe (NonEmpty Text)

The targets of the OTA update.

otaUpdateFiles :: Maybe (NonEmpty OTAUpdateFile)

A list of files associated with the OTA update.

awsJobPresignedUrlConfig :: Maybe AwsJobPresignedUrlConfig

Configuration information for pre-signed URLs. Valid when protocols contains HTTP.

errorInfo :: Maybe ErrorInfo

Error information associated with the OTA update.

otaUpdateId :: Maybe Text

The OTA update ID.

description :: Maybe Text

A description of the OTA update.

additionalParameters :: Maybe (HashMap Text Text)

A collection of name/value pairs

awsJobExecutionsRolloutConfig :: Maybe AwsJobExecutionsRolloutConfig

Configuration for the rollout of OTA updates.

newOTAUpdateInfo :: OTAUpdateInfo

Create a value of OTAUpdateInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:otaUpdateStatus:OTAUpdateInfo', oTAUpdateInfo_otaUpdateStatus - The status of the OTA update.

$sel:lastModifiedDate:OTAUpdateInfo', oTAUpdateInfo_lastModifiedDate - The date when the OTA update was last updated.

$sel:targetSelection:OTAUpdateInfo', oTAUpdateInfo_targetSelection - Specifies whether the OTA update will continue to run (CONTINUOUS), or will be complete after all those things specified as targets have completed the OTA update (SNAPSHOT). If continuous, the OTA update may also be run on a thing when a change is detected in a target. For example, an OTA update will run on a thing when the thing is added to a target group, even after the OTA update was completed by all things originally in the group.

$sel:otaUpdateArn:OTAUpdateInfo', oTAUpdateInfo_otaUpdateArn - The OTA update ARN.

$sel:creationDate:OTAUpdateInfo', oTAUpdateInfo_creationDate - The date when the OTA update was created.

$sel:protocols:OTAUpdateInfo', oTAUpdateInfo_protocols - The protocol used to transfer the OTA update image. Valid values are [HTTP], [MQTT], [HTTP, MQTT]. When both HTTP and MQTT are specified, the target device can choose the protocol.

$sel:awsIotJobArn:OTAUpdateInfo', oTAUpdateInfo_awsIotJobArn - The IoT job ARN associated with the OTA update.

$sel:awsIotJobId:OTAUpdateInfo', oTAUpdateInfo_awsIotJobId - The IoT job ID associated with the OTA update.

$sel:targets:OTAUpdateInfo', oTAUpdateInfo_targets - The targets of the OTA update.

$sel:otaUpdateFiles:OTAUpdateInfo', oTAUpdateInfo_otaUpdateFiles - A list of files associated with the OTA update.

$sel:awsJobPresignedUrlConfig:OTAUpdateInfo', oTAUpdateInfo_awsJobPresignedUrlConfig - Configuration information for pre-signed URLs. Valid when protocols contains HTTP.

$sel:errorInfo:OTAUpdateInfo', oTAUpdateInfo_errorInfo - Error information associated with the OTA update.

$sel:otaUpdateId:OTAUpdateInfo', oTAUpdateInfo_otaUpdateId - The OTA update ID.

$sel:description:OTAUpdateInfo', oTAUpdateInfo_description - A description of the OTA update.

$sel:additionalParameters:OTAUpdateInfo', oTAUpdateInfo_additionalParameters - A collection of name/value pairs

$sel:awsJobExecutionsRolloutConfig:OTAUpdateInfo', oTAUpdateInfo_awsJobExecutionsRolloutConfig - Configuration for the rollout of OTA updates.

oTAUpdateInfo_otaUpdateStatus :: Lens' OTAUpdateInfo (Maybe OTAUpdateStatus)

The status of the OTA update.

oTAUpdateInfo_lastModifiedDate :: Lens' OTAUpdateInfo (Maybe UTCTime)

The date when the OTA update was last updated.

oTAUpdateInfo_targetSelection :: Lens' OTAUpdateInfo (Maybe TargetSelection)

Specifies whether the OTA update will continue to run (CONTINUOUS), or will be complete after all those things specified as targets have completed the OTA update (SNAPSHOT). If continuous, the OTA update may also be run on a thing when a change is detected in a target. For example, an OTA update will run on a thing when the thing is added to a target group, even after the OTA update was completed by all things originally in the group.

oTAUpdateInfo_otaUpdateArn :: Lens' OTAUpdateInfo (Maybe Text)

The OTA update ARN.

oTAUpdateInfo_creationDate :: Lens' OTAUpdateInfo (Maybe UTCTime)

The date when the OTA update was created.

oTAUpdateInfo_protocols :: Lens' OTAUpdateInfo (Maybe (NonEmpty Protocol))

The protocol used to transfer the OTA update image. Valid values are [HTTP], [MQTT], [HTTP, MQTT]. When both HTTP and MQTT are specified, the target device can choose the protocol.

oTAUpdateInfo_awsIotJobArn :: Lens' OTAUpdateInfo (Maybe Text)

The IoT job ARN associated with the OTA update.

oTAUpdateInfo_awsIotJobId :: Lens' OTAUpdateInfo (Maybe Text)

The IoT job ID associated with the OTA update.

oTAUpdateInfo_targets :: Lens' OTAUpdateInfo (Maybe (NonEmpty Text))

The targets of the OTA update.

oTAUpdateInfo_otaUpdateFiles :: Lens' OTAUpdateInfo (Maybe (NonEmpty OTAUpdateFile))

A list of files associated with the OTA update.

oTAUpdateInfo_awsJobPresignedUrlConfig :: Lens' OTAUpdateInfo (Maybe AwsJobPresignedUrlConfig)

Configuration information for pre-signed URLs. Valid when protocols contains HTTP.

oTAUpdateInfo_errorInfo :: Lens' OTAUpdateInfo (Maybe ErrorInfo)

Error information associated with the OTA update.

oTAUpdateInfo_otaUpdateId :: Lens' OTAUpdateInfo (Maybe Text)

The OTA update ID.

oTAUpdateInfo_description :: Lens' OTAUpdateInfo (Maybe Text)

A description of the OTA update.

oTAUpdateInfo_additionalParameters :: Lens' OTAUpdateInfo (Maybe (HashMap Text Text))

A collection of name/value pairs

oTAUpdateInfo_awsJobExecutionsRolloutConfig :: Lens' OTAUpdateInfo (Maybe AwsJobExecutionsRolloutConfig)

Configuration for the rollout of OTA updates.