Network.AWS.IoT.Types.PutAssetPropertyValueEntry

Description

Documentation

data PutAssetPropertyValueEntry

An asset property value entry containing the following information.

See: newPutAssetPropertyValueEntry smart constructor.

Constructors

PutAssetPropertyValueEntry'

Fields

propertyAlias :: Maybe Text

The name of the property alias associated with your asset property. You must specify either a propertyAlias or both an aliasId and a propertyId. Accepts substitution templates.

entryId :: Maybe Text

Optional. A unique identifier for this entry that you can define to better track which message caused an error in case of failure. Accepts substitution templates. Defaults to a new UUID.

assetId :: Maybe Text

The ID of the IoT SiteWise asset. You must specify either a propertyAlias or both an aliasId and a propertyId. Accepts substitution templates.

propertyId :: Maybe Text

The ID of the asset's property. You must specify either a propertyAlias or both an aliasId and a propertyId. Accepts substitution templates.

propertyValues :: NonEmpty AssetPropertyValue

A list of property values to insert that each contain timestamp, quality, and value (TQV) information.

newPutAssetPropertyValueEntry

Arguments

:: NonEmpty AssetPropertyValue
$sel:propertyValues:PutAssetPropertyValueEntry'

-> PutAssetPropertyValueEntry

Create a value of PutAssetPropertyValueEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:propertyAlias:PutAssetPropertyValueEntry', putAssetPropertyValueEntry_propertyAlias - The name of the property alias associated with your asset property. You must specify either a propertyAlias or both an aliasId and a propertyId. Accepts substitution templates.

$sel:entryId:PutAssetPropertyValueEntry', putAssetPropertyValueEntry_entryId - Optional. A unique identifier for this entry that you can define to better track which message caused an error in case of failure. Accepts substitution templates. Defaults to a new UUID.

$sel:assetId:PutAssetPropertyValueEntry', putAssetPropertyValueEntry_assetId - The ID of the IoT SiteWise asset. You must specify either a propertyAlias or both an aliasId and a propertyId. Accepts substitution templates.

$sel:propertyId:PutAssetPropertyValueEntry', putAssetPropertyValueEntry_propertyId - The ID of the asset's property. You must specify either a propertyAlias or both an aliasId and a propertyId. Accepts substitution templates.

$sel:propertyValues:PutAssetPropertyValueEntry', putAssetPropertyValueEntry_propertyValues - A list of property values to insert that each contain timestamp, quality, and value (TQV) information.

putAssetPropertyValueEntry_propertyAlias :: Lens' PutAssetPropertyValueEntry (Maybe Text)

The name of the property alias associated with your asset property. You must specify either a propertyAlias or both an aliasId and a propertyId. Accepts substitution templates.

putAssetPropertyValueEntry_entryId :: Lens' PutAssetPropertyValueEntry (Maybe Text)

Optional. A unique identifier for this entry that you can define to better track which message caused an error in case of failure. Accepts substitution templates. Defaults to a new UUID.

putAssetPropertyValueEntry_assetId :: Lens' PutAssetPropertyValueEntry (Maybe Text)

The ID of the IoT SiteWise asset. You must specify either a propertyAlias or both an aliasId and a propertyId. Accepts substitution templates.

putAssetPropertyValueEntry_propertyId :: Lens' PutAssetPropertyValueEntry (Maybe Text)

The ID of the asset's property. You must specify either a propertyAlias or both an aliasId and a propertyId. Accepts substitution templates.

putAssetPropertyValueEntry_propertyValues :: Lens' PutAssetPropertyValueEntry (NonEmpty AssetPropertyValue)

A list of property values to insert that each contain timestamp, quality, and value (TQV) information.