Network.AWS.IoT.Types.ScheduledAuditMetadata

Description

Documentation

data ScheduledAuditMetadata

Information about the scheduled audit.

See: newScheduledAuditMetadata smart constructor.

Constructors

ScheduledAuditMetadata'

Fields

dayOfWeek :: Maybe DayOfWeek

The day of the week on which the scheduled audit is run (if the frequency is "WEEKLY" or "BIWEEKLY").

scheduledAuditArn :: Maybe Text

The ARN of the scheduled audit.

scheduledAuditName :: Maybe Text

The name of the scheduled audit.

dayOfMonth :: Maybe Text

The day of the month on which the scheduled audit is run (if the frequency is "MONTHLY"). If days 29-31 are specified, and the month does not have that many days, the audit takes place on the "LAST" day of the month.

frequency :: Maybe AuditFrequency

How often the scheduled audit occurs.

newScheduledAuditMetadata :: ScheduledAuditMetadata

Create a value of ScheduledAuditMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dayOfWeek:ScheduledAuditMetadata', scheduledAuditMetadata_dayOfWeek - The day of the week on which the scheduled audit is run (if the frequency is "WEEKLY" or "BIWEEKLY").

$sel:scheduledAuditArn:ScheduledAuditMetadata', scheduledAuditMetadata_scheduledAuditArn - The ARN of the scheduled audit.

$sel:scheduledAuditName:ScheduledAuditMetadata', scheduledAuditMetadata_scheduledAuditName - The name of the scheduled audit.

$sel:dayOfMonth:ScheduledAuditMetadata', scheduledAuditMetadata_dayOfMonth - The day of the month on which the scheduled audit is run (if the frequency is "MONTHLY"). If days 29-31 are specified, and the month does not have that many days, the audit takes place on the "LAST" day of the month.

$sel:frequency:ScheduledAuditMetadata', scheduledAuditMetadata_frequency - How often the scheduled audit occurs.

scheduledAuditMetadata_dayOfWeek :: Lens' ScheduledAuditMetadata (Maybe DayOfWeek)

The day of the week on which the scheduled audit is run (if the frequency is "WEEKLY" or "BIWEEKLY").

scheduledAuditMetadata_scheduledAuditArn :: Lens' ScheduledAuditMetadata (Maybe Text)

The ARN of the scheduled audit.

scheduledAuditMetadata_scheduledAuditName :: Lens' ScheduledAuditMetadata (Maybe Text)

The name of the scheduled audit.

scheduledAuditMetadata_dayOfMonth :: Lens' ScheduledAuditMetadata (Maybe Text)

The day of the month on which the scheduled audit is run (if the frequency is "MONTHLY"). If days 29-31 are specified, and the month does not have that many days, the audit takes place on the "LAST" day of the month.

scheduledAuditMetadata_frequency :: Lens' ScheduledAuditMetadata (Maybe AuditFrequency)

How often the scheduled audit occurs.