Network.AWS.IoT.Types.ServerCertificateSummary

Description

Documentation

data ServerCertificateSummary

An object that contains information about a server certificate.

See: newServerCertificateSummary smart constructor.

Constructors

ServerCertificateSummary'

Fields

serverCertificateStatus :: Maybe ServerCertificateStatus

The status of the server certificate.

serverCertificateStatusDetail :: Maybe Text

Details that explain the status of the server certificate.

serverCertificateArn :: Maybe Text

The ARN of the server certificate.

newServerCertificateSummary :: ServerCertificateSummary

Create a value of ServerCertificateSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serverCertificateStatus:ServerCertificateSummary', serverCertificateSummary_serverCertificateStatus - The status of the server certificate.

$sel:serverCertificateStatusDetail:ServerCertificateSummary', serverCertificateSummary_serverCertificateStatusDetail - Details that explain the status of the server certificate.

$sel:serverCertificateArn:ServerCertificateSummary', serverCertificateSummary_serverCertificateArn - The ARN of the server certificate.

serverCertificateSummary_serverCertificateStatus :: Lens' ServerCertificateSummary (Maybe ServerCertificateStatus)

The status of the server certificate.

serverCertificateSummary_serverCertificateStatusDetail :: Lens' ServerCertificateSummary (Maybe Text)

Details that explain the status of the server certificate.

serverCertificateSummary_serverCertificateArn :: Lens' ServerCertificateSummary (Maybe Text)

The ARN of the server certificate.