Network.AWS.IoT.Types.SnsAction

Description

Documentation

data SnsAction

Describes an action to publish to an Amazon SNS topic.

See: newSnsAction smart constructor.

Constructors

SnsAction'

Fields

messageFormat :: Maybe MessageFormat

(Optional) The message format of the message to publish. Accepted values are "JSON" and "RAW". The default value of the attribute is "RAW". SNS uses this setting to determine if the payload should be parsed and relevant platform-specific bits of the payload should be extracted. To read more about SNS message formats, see https://docs.aws.amazon.com/sns/latest/dg/json-formats.html refer to their official documentation.

targetArn :: Text

The ARN of the SNS topic.

roleArn :: Text

The ARN of the IAM role that grants access.

newSnsAction

Arguments

:: Text
$sel:targetArn:SnsAction'

-> Text
$sel:roleArn:SnsAction'

-> SnsAction

Create a value of SnsAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:messageFormat:SnsAction', snsAction_messageFormat - (Optional) The message format of the message to publish. Accepted values are "JSON" and "RAW". The default value of the attribute is "RAW". SNS uses this setting to determine if the payload should be parsed and relevant platform-specific bits of the payload should be extracted. To read more about SNS message formats, see https://docs.aws.amazon.com/sns/latest/dg/json-formats.html refer to their official documentation.

$sel:targetArn:SnsAction', snsAction_targetArn - The ARN of the SNS topic.

$sel:roleArn:SnsAction', snsAction_roleArn - The ARN of the IAM role that grants access.

snsAction_messageFormat :: Lens' SnsAction (Maybe MessageFormat)

(Optional) The message format of the message to publish. Accepted values are "JSON" and "RAW". The default value of the attribute is "RAW". SNS uses this setting to determine if the payload should be parsed and relevant platform-specific bits of the payload should be extracted. To read more about SNS message formats, see https://docs.aws.amazon.com/sns/latest/dg/json-formats.html refer to their official documentation.

snsAction_targetArn :: Lens' SnsAction Text

The ARN of the SNS topic.

snsAction_roleArn :: Lens' SnsAction Text

The ARN of the IAM role that grants access.