Network.AWS.IoT.Types.Statistics

Description

Documentation

data Statistics

A map of key-value pairs for all supported statistics. Currently, only count is supported.

See: newStatistics smart constructor.

Constructors

Statistics'

Fields

minimum :: Maybe Double

The minimum aggregated field value.

sum :: Maybe Double

The sum of the aggregated field values.

stdDeviation :: Maybe Double

The standard deviation of the aggregated field values.

variance :: Maybe Double

The variance of the aggregated field values.

average :: Maybe Double

The average of the aggregated field values.

count :: Maybe Int

The count of things that match the query.

sumOfSquares :: Maybe Double

The sum of the squares of the aggregated field values.

maximum :: Maybe Double

The maximum aggregated field value.

newStatistics :: Statistics

Create a value of Statistics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:minimum:Statistics', statistics_minimum - The minimum aggregated field value.

$sel:sum:Statistics', statistics_sum - The sum of the aggregated field values.

$sel:stdDeviation:Statistics', statistics_stdDeviation - The standard deviation of the aggregated field values.

$sel:variance:Statistics', statistics_variance - The variance of the aggregated field values.

$sel:average:Statistics', statistics_average - The average of the aggregated field values.

$sel:count:Statistics', statistics_count - The count of things that match the query.

$sel:sumOfSquares:Statistics', statistics_sumOfSquares - The sum of the squares of the aggregated field values.

$sel:maximum:Statistics', statistics_maximum - The maximum aggregated field value.

statistics_minimum :: Lens' Statistics (Maybe Double)

The minimum aggregated field value.

statistics_sum :: Lens' Statistics (Maybe Double)

The sum of the aggregated field values.

statistics_stdDeviation :: Lens' Statistics (Maybe Double)

The standard deviation of the aggregated field values.

statistics_variance :: Lens' Statistics (Maybe Double)

The variance of the aggregated field values.

statistics_average :: Lens' Statistics (Maybe Double)

The average of the aggregated field values.

statistics_count :: Lens' Statistics (Maybe Int)

The count of things that match the query.

statistics_sumOfSquares :: Lens' Statistics (Maybe Double)

The sum of the squares of the aggregated field values.

statistics_maximum :: Lens' Statistics (Maybe Double)

The maximum aggregated field value.