Network.AWS.IoT.Types.StepFunctionsAction

Description

Documentation

data StepFunctionsAction

Starts execution of a Step Functions state machine.

See: newStepFunctionsAction smart constructor.

Constructors

StepFunctionsAction'

Fields

executionNamePrefix :: Maybe Text

(Optional) A name will be given to the state machine execution consisting of this prefix followed by a UUID. Step Functions automatically creates a unique name for each state machine execution if one is not provided.

stateMachineName :: Text

The name of the Step Functions state machine whose execution will be started.

roleArn :: Text

The ARN of the role that grants IoT permission to start execution of a state machine ("Action":"states:StartExecution").

newStepFunctionsAction

Arguments

:: Text
$sel:stateMachineName:StepFunctionsAction'

-> Text
$sel:roleArn:StepFunctionsAction'

-> StepFunctionsAction

Create a value of StepFunctionsAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:executionNamePrefix:StepFunctionsAction', stepFunctionsAction_executionNamePrefix - (Optional) A name will be given to the state machine execution consisting of this prefix followed by a UUID. Step Functions automatically creates a unique name for each state machine execution if one is not provided.

$sel:stateMachineName:StepFunctionsAction', stepFunctionsAction_stateMachineName - The name of the Step Functions state machine whose execution will be started.

$sel:roleArn:StepFunctionsAction', stepFunctionsAction_roleArn - The ARN of the role that grants IoT permission to start execution of a state machine ("Action":"states:StartExecution").

stepFunctionsAction_executionNamePrefix :: Lens' StepFunctionsAction (Maybe Text)

(Optional) A name will be given to the state machine execution consisting of this prefix followed by a UUID. Step Functions automatically creates a unique name for each state machine execution if one is not provided.

stepFunctionsAction_stateMachineName :: Lens' StepFunctionsAction Text

The name of the Step Functions state machine whose execution will be started.

stepFunctionsAction_roleArn :: Lens' StepFunctionsAction Text

The ARN of the role that grants IoT permission to start execution of a state machine ("Action":"states:StartExecution").