Network.AWS.IoT.Types.TaskStatisticsForAuditCheck

Description

Documentation

data TaskStatisticsForAuditCheck

Provides summary counts of how many tasks for findings are in a particular state. This information is included in the response from DescribeAuditMitigationActionsTask.

See: newTaskStatisticsForAuditCheck smart constructor.

Constructors

TaskStatisticsForAuditCheck'

Fields

succeededFindingsCount :: Maybe Integer

The number of findings for which all mitigation actions succeeded when applied.

totalFindingsCount :: Maybe Integer

The total number of findings to which a task is being applied.

failedFindingsCount :: Maybe Integer

The number of findings for which at least one of the actions failed when applied.

skippedFindingsCount :: Maybe Integer

The number of findings skipped because of filter conditions provided in the parameters to the command.

canceledFindingsCount :: Maybe Integer

The number of findings to which the mitigation action task was canceled when applied.

newTaskStatisticsForAuditCheck :: TaskStatisticsForAuditCheck

Create a value of TaskStatisticsForAuditCheck with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:succeededFindingsCount:TaskStatisticsForAuditCheck', taskStatisticsForAuditCheck_succeededFindingsCount - The number of findings for which all mitigation actions succeeded when applied.

$sel:totalFindingsCount:TaskStatisticsForAuditCheck', taskStatisticsForAuditCheck_totalFindingsCount - The total number of findings to which a task is being applied.

$sel:failedFindingsCount:TaskStatisticsForAuditCheck', taskStatisticsForAuditCheck_failedFindingsCount - The number of findings for which at least one of the actions failed when applied.

$sel:skippedFindingsCount:TaskStatisticsForAuditCheck', taskStatisticsForAuditCheck_skippedFindingsCount - The number of findings skipped because of filter conditions provided in the parameters to the command.

$sel:canceledFindingsCount:TaskStatisticsForAuditCheck', taskStatisticsForAuditCheck_canceledFindingsCount - The number of findings to which the mitigation action task was canceled when applied.

taskStatisticsForAuditCheck_succeededFindingsCount :: Lens' TaskStatisticsForAuditCheck (Maybe Integer)

The number of findings for which all mitigation actions succeeded when applied.

taskStatisticsForAuditCheck_totalFindingsCount :: Lens' TaskStatisticsForAuditCheck (Maybe Integer)

The total number of findings to which a task is being applied.

taskStatisticsForAuditCheck_failedFindingsCount :: Lens' TaskStatisticsForAuditCheck (Maybe Integer)

The number of findings for which at least one of the actions failed when applied.

taskStatisticsForAuditCheck_skippedFindingsCount :: Lens' TaskStatisticsForAuditCheck (Maybe Integer)

The number of findings skipped because of filter conditions provided in the parameters to the command.

taskStatisticsForAuditCheck_canceledFindingsCount :: Lens' TaskStatisticsForAuditCheck (Maybe Integer)

The number of findings to which the mitigation action task was canceled when applied.