Network.AWS.IoT.Types.ThingAttribute

Description

Documentation

data ThingAttribute

The properties of the thing, including thing name, thing type name, and a list of thing attributes.

See: newThingAttribute smart constructor.

Constructors

ThingAttribute'

Fields

thingArn :: Maybe Text

The thing ARN.

thingName :: Maybe Text

The name of the thing.

version :: Maybe Integer

The version of the thing record in the registry.

attributes :: Maybe (HashMap Text Text)

A list of thing attributes which are name-value pairs.

thingTypeName :: Maybe Text

The name of the thing type, if the thing has been associated with a type.

newThingAttribute :: ThingAttribute

Create a value of ThingAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thingArn:ThingAttribute', thingAttribute_thingArn - The thing ARN.

$sel:thingName:ThingAttribute', thingAttribute_thingName - The name of the thing.

$sel:version:ThingAttribute', thingAttribute_version - The version of the thing record in the registry.

$sel:attributes:ThingAttribute', thingAttribute_attributes - A list of thing attributes which are name-value pairs.

$sel:thingTypeName:ThingAttribute', thingAttribute_thingTypeName - The name of the thing type, if the thing has been associated with a type.

thingAttribute_thingArn :: Lens' ThingAttribute (Maybe Text)

The thing ARN.

thingAttribute_thingName :: Lens' ThingAttribute (Maybe Text)

The name of the thing.

thingAttribute_version :: Lens' ThingAttribute (Maybe Integer)

The version of the thing record in the registry.

thingAttribute_attributes :: Lens' ThingAttribute (Maybe (HashMap Text Text))

A list of thing attributes which are name-value pairs.

thingAttribute_thingTypeName :: Lens' ThingAttribute (Maybe Text)

The name of the thing type, if the thing has been associated with a type.