Network.AWS.IoT.Types.ThingConnectivity

Description

Documentation

data ThingConnectivity

The connectivity status of the thing.

See: newThingConnectivity smart constructor.

Constructors

ThingConnectivity'

Fields

disconnectReason :: Maybe Text

The reason why the client is disconnected. If the thing has been disconnected for approximately an hour, the disconnectReason value might be missing.

connected :: Maybe Bool

True if the thing is connected to the Amazon Web Services IoT Core service; false if it is not connected.

timestamp :: Maybe Integer

The epoch time (in milliseconds) when the thing last connected or disconnected. If the thing has been disconnected for approximately an hour, the time value might be missing.

newThingConnectivity :: ThingConnectivity

Create a value of ThingConnectivity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:disconnectReason:ThingConnectivity', thingConnectivity_disconnectReason - The reason why the client is disconnected. If the thing has been disconnected for approximately an hour, the disconnectReason value might be missing.

$sel:connected:ThingConnectivity', thingConnectivity_connected - True if the thing is connected to the Amazon Web Services IoT Core service; false if it is not connected.

$sel:timestamp:ThingConnectivity', thingConnectivity_timestamp - The epoch time (in milliseconds) when the thing last connected or disconnected. If the thing has been disconnected for approximately an hour, the time value might be missing.

thingConnectivity_disconnectReason :: Lens' ThingConnectivity (Maybe Text)

The reason why the client is disconnected. If the thing has been disconnected for approximately an hour, the disconnectReason value might be missing.

thingConnectivity_connected :: Lens' ThingConnectivity (Maybe Bool)

True if the thing is connected to the Amazon Web Services IoT Core service; false if it is not connected.

thingConnectivity_timestamp :: Lens' ThingConnectivity (Maybe Integer)

The epoch time (in milliseconds) when the thing last connected or disconnected. If the thing has been disconnected for approximately an hour, the time value might be missing.