Network.AWS.IoT.Types.ThingTypeDefinition

Description

Documentation

data ThingTypeDefinition

The definition of the thing type, including thing type name and description.

See: newThingTypeDefinition smart constructor.

Constructors

ThingTypeDefinition'

Fields

thingTypeProperties :: Maybe ThingTypeProperties

The ThingTypeProperties for the thing type.

thingTypeName :: Maybe Text

The name of the thing type.

thingTypeMetadata :: Maybe ThingTypeMetadata

The ThingTypeMetadata contains additional information about the thing type including: creation date and time, a value indicating whether the thing type is deprecated, and a date and time when it was deprecated.

thingTypeArn :: Maybe Text

The thing type ARN.

newThingTypeDefinition :: ThingTypeDefinition

Create a value of ThingTypeDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thingTypeProperties:ThingTypeDefinition', thingTypeDefinition_thingTypeProperties - The ThingTypeProperties for the thing type.

$sel:thingTypeName:ThingTypeDefinition', thingTypeDefinition_thingTypeName - The name of the thing type.

$sel:thingTypeMetadata:ThingTypeDefinition', thingTypeDefinition_thingTypeMetadata - The ThingTypeMetadata contains additional information about the thing type including: creation date and time, a value indicating whether the thing type is deprecated, and a date and time when it was deprecated.

$sel:thingTypeArn:ThingTypeDefinition', thingTypeDefinition_thingTypeArn - The thing type ARN.

thingTypeDefinition_thingTypeProperties :: Lens' ThingTypeDefinition (Maybe ThingTypeProperties)

The ThingTypeProperties for the thing type.

thingTypeDefinition_thingTypeName :: Lens' ThingTypeDefinition (Maybe Text)

The name of the thing type.

thingTypeDefinition_thingTypeMetadata :: Lens' ThingTypeDefinition (Maybe ThingTypeMetadata)

The ThingTypeMetadata contains additional information about the thing type including: creation date and time, a value indicating whether the thing type is deprecated, and a date and time when it was deprecated.

thingTypeDefinition_thingTypeArn :: Lens' ThingTypeDefinition (Maybe Text)

The thing type ARN.