Network.AWS.IoT.Types.TopicRulePayload

Description

Documentation

data TopicRulePayload

Describes a rule.

See: newTopicRulePayload smart constructor.

Constructors

TopicRulePayload'

Fields

errorAction :: Maybe Action

The action to take when an error occurs.

awsIotSqlVersion :: Maybe Text

The version of the SQL rules engine to use when evaluating the rule.

ruleDisabled :: Maybe Bool

Specifies whether the rule is disabled.

description :: Maybe Text

The description of the rule.

sql :: Text

The SQL statement used to query the topic. For more information, see IoT SQL Reference in the IoT Developer Guide.

actions :: [Action]

The actions associated with the rule.

newTopicRulePayload

Arguments

:: Text
$sel:sql:TopicRulePayload'

-> TopicRulePayload

Create a value of TopicRulePayload with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorAction:TopicRulePayload', topicRulePayload_errorAction - The action to take when an error occurs.

$sel:awsIotSqlVersion:TopicRulePayload', topicRulePayload_awsIotSqlVersion - The version of the SQL rules engine to use when evaluating the rule.

$sel:ruleDisabled:TopicRulePayload', topicRulePayload_ruleDisabled - Specifies whether the rule is disabled.

$sel:description:TopicRulePayload', topicRulePayload_description - The description of the rule.

$sel:sql:TopicRulePayload', topicRulePayload_sql - The SQL statement used to query the topic. For more information, see IoT SQL Reference in the IoT Developer Guide.

$sel:actions:TopicRulePayload', topicRulePayload_actions - The actions associated with the rule.

topicRulePayload_errorAction :: Lens' TopicRulePayload (Maybe Action)

The action to take when an error occurs.

topicRulePayload_awsIotSqlVersion :: Lens' TopicRulePayload (Maybe Text)

The version of the SQL rules engine to use when evaluating the rule.

topicRulePayload_ruleDisabled :: Lens' TopicRulePayload (Maybe Bool)

Specifies whether the rule is disabled.

topicRulePayload_description :: Lens' TopicRulePayload (Maybe Text)

The description of the rule.

topicRulePayload_sql :: Lens' TopicRulePayload Text

The SQL statement used to query the topic. For more information, see IoT SQL Reference in the IoT Developer Guide.

topicRulePayload_actions :: Lens' TopicRulePayload [Action]

The actions associated with the rule.