Network.AWS.IoT.Types.ViolationEvent

Description

Documentation

data ViolationEvent

Information about a Device Defender security profile behavior violation.

See: newViolationEvent smart constructor.

Constructors

ViolationEvent'

Fields

metricValue :: Maybe MetricValue

The value of the metric (the measurement).

violationId :: Maybe Text

The ID of the violation event.

thingName :: Maybe Text

The name of the thing responsible for the violation event.

verificationStateDescription :: Maybe Text

The description of the verification state of the violation.

securityProfileName :: Maybe Text

The name of the security profile whose behavior was violated.

behavior :: Maybe Behavior

The behavior that was violated.

verificationState :: Maybe VerificationState

The verification state of the violation (detect alarm).

violationEventTime :: Maybe POSIX

The time the violation event occurred.

violationEventType :: Maybe ViolationEventType

The type of violation event.

violationEventAdditionalInfo :: Maybe ViolationEventAdditionalInfo

The details of a violation event.

newViolationEvent :: ViolationEvent

Create a value of ViolationEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricValue:ViolationEvent', violationEvent_metricValue - The value of the metric (the measurement).

$sel:violationId:ViolationEvent', violationEvent_violationId - The ID of the violation event.

$sel:thingName:ViolationEvent', violationEvent_thingName - The name of the thing responsible for the violation event.

$sel:verificationStateDescription:ViolationEvent', violationEvent_verificationStateDescription - The description of the verification state of the violation.

$sel:securityProfileName:ViolationEvent', violationEvent_securityProfileName - The name of the security profile whose behavior was violated.

$sel:behavior:ViolationEvent', violationEvent_behavior - The behavior that was violated.

$sel:verificationState:ViolationEvent', violationEvent_verificationState - The verification state of the violation (detect alarm).

$sel:violationEventTime:ViolationEvent', violationEvent_violationEventTime - The time the violation event occurred.

$sel:violationEventType:ViolationEvent', violationEvent_violationEventType - The type of violation event.

$sel:violationEventAdditionalInfo:ViolationEvent', violationEvent_violationEventAdditionalInfo - The details of a violation event.

violationEvent_metricValue :: Lens' ViolationEvent (Maybe MetricValue)

The value of the metric (the measurement).

violationEvent_violationId :: Lens' ViolationEvent (Maybe Text)

The ID of the violation event.

violationEvent_thingName :: Lens' ViolationEvent (Maybe Text)

The name of the thing responsible for the violation event.

violationEvent_verificationStateDescription :: Lens' ViolationEvent (Maybe Text)

The description of the verification state of the violation.

violationEvent_securityProfileName :: Lens' ViolationEvent (Maybe Text)

The name of the security profile whose behavior was violated.

violationEvent_behavior :: Lens' ViolationEvent (Maybe Behavior)

The behavior that was violated.

violationEvent_verificationState :: Lens' ViolationEvent (Maybe VerificationState)

The verification state of the violation (detect alarm).

violationEvent_violationEventTime :: Lens' ViolationEvent (Maybe UTCTime)

The time the violation event occurred.

violationEvent_violationEventType :: Lens' ViolationEvent (Maybe ViolationEventType)

The type of violation event.

violationEvent_violationEventAdditionalInfo :: Lens' ViolationEvent (Maybe ViolationEventAdditionalInfo)

The details of a violation event.