Network.AWS.IoT.UpdateAuthorizer

Description

Updates an authorizer.

Requires permission to access the UpdateAuthorizer action.

Creating a Request

data UpdateAuthorizer

See: newUpdateAuthorizer smart constructor.

Constructors

UpdateAuthorizer'

Fields

status :: Maybe AuthorizerStatus

The status of the update authorizer request.

authorizerFunctionArn :: Maybe Text

The ARN of the authorizer's Lambda function.

tokenSigningPublicKeys :: Maybe (HashMap Text Text)

The public keys used to verify the token signature.

tokenKeyName :: Maybe Text

The key used to extract the token from the HTTP headers.

authorizerName :: Text

The authorizer name.

newUpdateAuthorizer

Arguments

:: Text
$sel:authorizerName:UpdateAuthorizer'

-> UpdateAuthorizer

Create a value of UpdateAuthorizer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UpdateAuthorizer', updateAuthorizer_status - The status of the update authorizer request.

$sel:authorizerFunctionArn:UpdateAuthorizer', updateAuthorizer_authorizerFunctionArn - The ARN of the authorizer's Lambda function.

$sel:tokenSigningPublicKeys:UpdateAuthorizer', updateAuthorizer_tokenSigningPublicKeys - The public keys used to verify the token signature.

$sel:tokenKeyName:UpdateAuthorizer', updateAuthorizer_tokenKeyName - The key used to extract the token from the HTTP headers.

$sel:authorizerName:UpdateAuthorizer', updateAuthorizer_authorizerName - The authorizer name.

Request Lenses

updateAuthorizer_status :: Lens' UpdateAuthorizer (Maybe AuthorizerStatus)

The status of the update authorizer request.

updateAuthorizer_authorizerFunctionArn :: Lens' UpdateAuthorizer (Maybe Text)

The ARN of the authorizer's Lambda function.

updateAuthorizer_tokenSigningPublicKeys :: Lens' UpdateAuthorizer (Maybe (HashMap Text Text))

The public keys used to verify the token signature.

updateAuthorizer_tokenKeyName :: Lens' UpdateAuthorizer (Maybe Text)

The key used to extract the token from the HTTP headers.

updateAuthorizer_authorizerName :: Lens' UpdateAuthorizer Text

The authorizer name.

Destructuring the Response

data UpdateAuthorizerResponse

See: newUpdateAuthorizerResponse smart constructor.

Constructors

UpdateAuthorizerResponse'

Fields

authorizerName :: Maybe Text

The authorizer name.

authorizerArn :: Maybe Text

The authorizer ARN.

httpStatus :: Int

The response's http status code.

newUpdateAuthorizerResponse

Arguments

:: Int
$sel:httpStatus:UpdateAuthorizerResponse'

-> UpdateAuthorizerResponse

Create a value of UpdateAuthorizerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authorizerName:UpdateAuthorizer', updateAuthorizerResponse_authorizerName - The authorizer name.

$sel:authorizerArn:UpdateAuthorizerResponse', updateAuthorizerResponse_authorizerArn - The authorizer ARN.

$sel:httpStatus:UpdateAuthorizerResponse', updateAuthorizerResponse_httpStatus - The response's http status code.

Response Lenses

updateAuthorizerResponse_authorizerName :: Lens' UpdateAuthorizerResponse (Maybe Text)

The authorizer name.

updateAuthorizerResponse_authorizerArn :: Lens' UpdateAuthorizerResponse (Maybe Text)

The authorizer ARN.

updateAuthorizerResponse_httpStatus :: Lens' UpdateAuthorizerResponse Int

The response's http status code.