Network.AWS.IoT.UpdateDimension

Description

Updates the definition for a dimension. You cannot change the type of a dimension after it is created (you can delete it and recreate it).

Requires permission to access the UpdateDimension action.

Creating a Request

data UpdateDimension

See: newUpdateDimension smart constructor.

Constructors

UpdateDimension'

Fields

name :: Text

A unique identifier for the dimension. Choose something that describes the type and value to make it easy to remember what it does.

stringValues :: NonEmpty Text

Specifies the value or list of values for the dimension. For TOPIC_FILTER dimensions, this is a pattern used to match the MQTT topic (for example, "admin/#").

newUpdateDimension

Arguments

:: Text
$sel:name:UpdateDimension'

-> NonEmpty Text
$sel:stringValues:UpdateDimension'

-> UpdateDimension

Create a value of UpdateDimension with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:UpdateDimension', updateDimension_name - A unique identifier for the dimension. Choose something that describes the type and value to make it easy to remember what it does.

$sel:stringValues:UpdateDimension', updateDimension_stringValues - Specifies the value or list of values for the dimension. For TOPIC_FILTER dimensions, this is a pattern used to match the MQTT topic (for example, "admin/#").

Request Lenses

updateDimension_name :: Lens' UpdateDimension Text

A unique identifier for the dimension. Choose something that describes the type and value to make it easy to remember what it does.

updateDimension_stringValues :: Lens' UpdateDimension (NonEmpty Text)

Specifies the value or list of values for the dimension. For TOPIC_FILTER dimensions, this is a pattern used to match the MQTT topic (for example, "admin/#").

Destructuring the Response

data UpdateDimensionResponse

See: newUpdateDimensionResponse smart constructor.

Constructors

UpdateDimensionResponse'

Fields

lastModifiedDate :: Maybe POSIX

The date and time, in milliseconds since epoch, when the dimension was most recently updated.

creationDate :: Maybe POSIX

The date and time, in milliseconds since epoch, when the dimension was initially created.

arn :: Maybe Text

The Amazon Resource Name (ARN)of the created dimension.

name :: Maybe Text

A unique identifier for the dimension.

type' :: Maybe DimensionType

The type of the dimension.

stringValues :: Maybe (NonEmpty Text)

The value or list of values used to scope the dimension. For example, for topic filters, this is the pattern used to match the MQTT topic name.

httpStatus :: Int

The response's http status code.

newUpdateDimensionResponse

Arguments

:: Int
$sel:httpStatus:UpdateDimensionResponse'

-> UpdateDimensionResponse

Create a value of UpdateDimensionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:UpdateDimensionResponse', updateDimensionResponse_lastModifiedDate - The date and time, in milliseconds since epoch, when the dimension was most recently updated.

$sel:creationDate:UpdateDimensionResponse', updateDimensionResponse_creationDate - The date and time, in milliseconds since epoch, when the dimension was initially created.

$sel:arn:UpdateDimensionResponse', updateDimensionResponse_arn - The Amazon Resource Name (ARN)of the created dimension.

$sel:name:UpdateDimension', updateDimensionResponse_name - A unique identifier for the dimension.

$sel:type':UpdateDimensionResponse', updateDimensionResponse_type - The type of the dimension.

$sel:stringValues:UpdateDimension', updateDimensionResponse_stringValues - The value or list of values used to scope the dimension. For example, for topic filters, this is the pattern used to match the MQTT topic name.

$sel:httpStatus:UpdateDimensionResponse', updateDimensionResponse_httpStatus - The response's http status code.

Response Lenses

updateDimensionResponse_lastModifiedDate :: Lens' UpdateDimensionResponse (Maybe UTCTime)

The date and time, in milliseconds since epoch, when the dimension was most recently updated.

updateDimensionResponse_creationDate :: Lens' UpdateDimensionResponse (Maybe UTCTime)

The date and time, in milliseconds since epoch, when the dimension was initially created.

updateDimensionResponse_arn :: Lens' UpdateDimensionResponse (Maybe Text)

The Amazon Resource Name (ARN)of the created dimension.

updateDimensionResponse_name :: Lens' UpdateDimensionResponse (Maybe Text)

A unique identifier for the dimension.

updateDimensionResponse_type :: Lens' UpdateDimensionResponse (Maybe DimensionType)

The type of the dimension.

updateDimensionResponse_stringValues :: Lens' UpdateDimensionResponse (Maybe (NonEmpty Text))

The value or list of values used to scope the dimension. For example, for topic filters, this is the pattern used to match the MQTT topic name.

updateDimensionResponse_httpStatus :: Lens' UpdateDimensionResponse Int

The response's http status code.