Network.AWS.IoT.UpdateDynamicThingGroup

Description

Updates a dynamic thing group.

Requires permission to access the UpdateDynamicThingGroup action.

Creating a Request

data UpdateDynamicThingGroup

See: newUpdateDynamicThingGroup smart constructor.

Constructors

UpdateDynamicThingGroup'

Fields

queryString :: Maybe Text

The dynamic thing group search query string to update.

expectedVersion :: Maybe Integer

The expected version of the dynamic thing group to update.

indexName :: Maybe Text

The dynamic thing group index to update.

Currently one index is supported: AWS_Things.

queryVersion :: Maybe Text

The dynamic thing group query version to update.

Currently one query version is supported: "2017-09-30". If not specified, the query version defaults to this value.

thingGroupName :: Text

The name of the dynamic thing group to update.

thingGroupProperties :: ThingGroupProperties

The dynamic thing group properties to update.

newUpdateDynamicThingGroup

Arguments

:: Text
$sel:thingGroupName:UpdateDynamicThingGroup'

-> ThingGroupProperties
$sel:thingGroupProperties:UpdateDynamicThingGroup'

-> UpdateDynamicThingGroup

Create a value of UpdateDynamicThingGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queryString:UpdateDynamicThingGroup', updateDynamicThingGroup_queryString - The dynamic thing group search query string to update.

$sel:expectedVersion:UpdateDynamicThingGroup', updateDynamicThingGroup_expectedVersion - The expected version of the dynamic thing group to update.

$sel:indexName:UpdateDynamicThingGroup', updateDynamicThingGroup_indexName - The dynamic thing group index to update.

Currently one index is supported: AWS_Things.

$sel:queryVersion:UpdateDynamicThingGroup', updateDynamicThingGroup_queryVersion - The dynamic thing group query version to update.

Currently one query version is supported: "2017-09-30". If not specified, the query version defaults to this value.

$sel:thingGroupName:UpdateDynamicThingGroup', updateDynamicThingGroup_thingGroupName - The name of the dynamic thing group to update.

$sel:thingGroupProperties:UpdateDynamicThingGroup', updateDynamicThingGroup_thingGroupProperties - The dynamic thing group properties to update.

Request Lenses

updateDynamicThingGroup_queryString :: Lens' UpdateDynamicThingGroup (Maybe Text)

The dynamic thing group search query string to update.

updateDynamicThingGroup_expectedVersion :: Lens' UpdateDynamicThingGroup (Maybe Integer)

The expected version of the dynamic thing group to update.

updateDynamicThingGroup_indexName :: Lens' UpdateDynamicThingGroup (Maybe Text)

The dynamic thing group index to update.

Currently one index is supported: AWS_Things.

updateDynamicThingGroup_queryVersion :: Lens' UpdateDynamicThingGroup (Maybe Text)

The dynamic thing group query version to update.

Currently one query version is supported: "2017-09-30". If not specified, the query version defaults to this value.

updateDynamicThingGroup_thingGroupName :: Lens' UpdateDynamicThingGroup Text

The name of the dynamic thing group to update.

updateDynamicThingGroup_thingGroupProperties :: Lens' UpdateDynamicThingGroup ThingGroupProperties

The dynamic thing group properties to update.

Destructuring the Response

data UpdateDynamicThingGroupResponse

See: newUpdateDynamicThingGroupResponse smart constructor.

Constructors

UpdateDynamicThingGroupResponse'

Fields

version :: Maybe Integer

The dynamic thing group version.

httpStatus :: Int

The response's http status code.

newUpdateDynamicThingGroupResponse

Arguments

:: Int
$sel:httpStatus:UpdateDynamicThingGroupResponse'

-> UpdateDynamicThingGroupResponse

Create a value of UpdateDynamicThingGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:version:UpdateDynamicThingGroupResponse', updateDynamicThingGroupResponse_version - The dynamic thing group version.

$sel:httpStatus:UpdateDynamicThingGroupResponse', updateDynamicThingGroupResponse_httpStatus - The response's http status code.

Response Lenses

updateDynamicThingGroupResponse_version :: Lens' UpdateDynamicThingGroupResponse (Maybe Integer)

The dynamic thing group version.

updateDynamicThingGroupResponse_httpStatus :: Lens' UpdateDynamicThingGroupResponse Int

The response's http status code.