Network.AWS.IoT.UpdateFleetMetric

Description

Updates the data for a fleet metric.

Requires permission to access the UpdateFleetMetric action.

Creating a Request

data UpdateFleetMetric

See: newUpdateFleetMetric smart constructor.

Constructors

UpdateFleetMetric'

Fields

queryString :: Maybe Text

The search query string.

expectedVersion :: Maybe Integer

The expected version of the fleet metric record in the registry.

unit :: Maybe FleetMetricUnit

Used to support unit transformation such as milliseconds to seconds. The unit must be supported by CW metric.

queryVersion :: Maybe Text

The version of the query.

aggregationType :: Maybe AggregationType

The type of the aggregation query.

aggregationField :: Maybe Text

The field to aggregate.

description :: Maybe Text

The description of the fleet metric.

period :: Maybe Natural

The time in seconds between fleet metric emissions. Range [60(1 min), 86400(1 day)] and must be multiple of 60.

metricName :: Text

The name of the fleet metric to update.

indexName :: Text

The name of the index to search.

newUpdateFleetMetric

Arguments

:: Text
$sel:metricName:UpdateFleetMetric'

-> Text
$sel:indexName:UpdateFleetMetric'

-> UpdateFleetMetric

Create a value of UpdateFleetMetric with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queryString:UpdateFleetMetric', updateFleetMetric_queryString - The search query string.

$sel:expectedVersion:UpdateFleetMetric', updateFleetMetric_expectedVersion - The expected version of the fleet metric record in the registry.

$sel:unit:UpdateFleetMetric', updateFleetMetric_unit - Used to support unit transformation such as milliseconds to seconds. The unit must be supported by CW metric.

$sel:queryVersion:UpdateFleetMetric', updateFleetMetric_queryVersion - The version of the query.

$sel:aggregationType:UpdateFleetMetric', updateFleetMetric_aggregationType - The type of the aggregation query.

$sel:aggregationField:UpdateFleetMetric', updateFleetMetric_aggregationField - The field to aggregate.

$sel:description:UpdateFleetMetric', updateFleetMetric_description - The description of the fleet metric.

$sel:period:UpdateFleetMetric', updateFleetMetric_period - The time in seconds between fleet metric emissions. Range [60(1 min), 86400(1 day)] and must be multiple of 60.

$sel:metricName:UpdateFleetMetric', updateFleetMetric_metricName - The name of the fleet metric to update.

$sel:indexName:UpdateFleetMetric', updateFleetMetric_indexName - The name of the index to search.

Request Lenses

updateFleetMetric_queryString :: Lens' UpdateFleetMetric (Maybe Text)

The search query string.

updateFleetMetric_expectedVersion :: Lens' UpdateFleetMetric (Maybe Integer)

The expected version of the fleet metric record in the registry.

updateFleetMetric_unit :: Lens' UpdateFleetMetric (Maybe FleetMetricUnit)

Used to support unit transformation such as milliseconds to seconds. The unit must be supported by CW metric.

updateFleetMetric_queryVersion :: Lens' UpdateFleetMetric (Maybe Text)

The version of the query.

updateFleetMetric_aggregationType :: Lens' UpdateFleetMetric (Maybe AggregationType)

The type of the aggregation query.

updateFleetMetric_aggregationField :: Lens' UpdateFleetMetric (Maybe Text)

The field to aggregate.

updateFleetMetric_description :: Lens' UpdateFleetMetric (Maybe Text)

The description of the fleet metric.

updateFleetMetric_period :: Lens' UpdateFleetMetric (Maybe Natural)

The time in seconds between fleet metric emissions. Range [60(1 min), 86400(1 day)] and must be multiple of 60.

updateFleetMetric_metricName :: Lens' UpdateFleetMetric Text

The name of the fleet metric to update.

updateFleetMetric_indexName :: Lens' UpdateFleetMetric Text

The name of the index to search.

Destructuring the Response

data UpdateFleetMetricResponse

See: newUpdateFleetMetricResponse smart constructor.

Constructors

UpdateFleetMetricResponse'

newUpdateFleetMetricResponse :: UpdateFleetMetricResponse

Create a value of UpdateFleetMetricResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.