Network.AWS.IoT.UpdateIndexingConfiguration

Description

Updates the search configuration.

Requires permission to access the UpdateIndexingConfiguration action.

Creating a Request

data UpdateIndexingConfiguration

See: newUpdateIndexingConfiguration smart constructor.

Constructors

UpdateIndexingConfiguration'

Fields

thingGroupIndexingConfiguration :: Maybe ThingGroupIndexingConfiguration

Thing group indexing configuration.

thingIndexingConfiguration :: Maybe ThingIndexingConfiguration

Thing indexing configuration.

newUpdateIndexingConfiguration :: UpdateIndexingConfiguration

Create a value of UpdateIndexingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thingGroupIndexingConfiguration:UpdateIndexingConfiguration', updateIndexingConfiguration_thingGroupIndexingConfiguration - Thing group indexing configuration.

$sel:thingIndexingConfiguration:UpdateIndexingConfiguration', updateIndexingConfiguration_thingIndexingConfiguration - Thing indexing configuration.

Request Lenses

updateIndexingConfiguration_thingGroupIndexingConfiguration :: Lens' UpdateIndexingConfiguration (Maybe ThingGroupIndexingConfiguration)

Thing group indexing configuration.

updateIndexingConfiguration_thingIndexingConfiguration :: Lens' UpdateIndexingConfiguration (Maybe ThingIndexingConfiguration)

Thing indexing configuration.

Destructuring the Response

data UpdateIndexingConfigurationResponse

See: newUpdateIndexingConfigurationResponse smart constructor.

Constructors

UpdateIndexingConfigurationResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateIndexingConfigurationResponse

Arguments

:: Int
$sel:httpStatus:UpdateIndexingConfigurationResponse'

-> UpdateIndexingConfigurationResponse

Create a value of UpdateIndexingConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateIndexingConfigurationResponse', updateIndexingConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

updateIndexingConfigurationResponse_httpStatus :: Lens' UpdateIndexingConfigurationResponse Int

The response's http status code.