Network.AWS.IoT.UpdateScheduledAudit

Description

Updates a scheduled audit, including which checks are performed and how often the audit takes place.

Requires permission to access the UpdateScheduledAudit action.

Creating a Request

data UpdateScheduledAudit

See: newUpdateScheduledAudit smart constructor.

Constructors

UpdateScheduledAudit'

Fields

frequency :: Maybe AuditFrequency

How often the scheduled audit takes place, either DAILY, WEEKLY, BIWEEKLY, or MONTHLY. The start time of each audit is determined by the system.

dayOfMonth :: Maybe Text

The day of the month on which the scheduled audit takes place. This can be 1 through 31 or LAST. This field is required if the frequency parameter is set to MONTHLY. If days 29-31 are specified, and the month does not have that many days, the audit takes place on the "LAST" day of the month.

targetCheckNames :: Maybe [Text]

Which checks are performed during the scheduled audit. Checks must be enabled for your account. (Use DescribeAccountAuditConfiguration to see the list of all checks, including those that are enabled or use UpdateAccountAuditConfiguration to select which checks are enabled.)

dayOfWeek :: Maybe DayOfWeek

The day of the week on which the scheduled audit takes place. This can be one of SUN, MON, TUE, WED, THU, FRI, or SAT. This field is required if the "frequency" parameter is set to WEEKLY or BIWEEKLY.

scheduledAuditName :: Text

The name of the scheduled audit. (Max. 128 chars)

newUpdateScheduledAudit

Arguments

:: Text
$sel:scheduledAuditName:UpdateScheduledAudit'

-> UpdateScheduledAudit

Create a value of UpdateScheduledAudit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:frequency:UpdateScheduledAudit', updateScheduledAudit_frequency - How often the scheduled audit takes place, either DAILY, WEEKLY, BIWEEKLY, or MONTHLY. The start time of each audit is determined by the system.

$sel:dayOfMonth:UpdateScheduledAudit', updateScheduledAudit_dayOfMonth - The day of the month on which the scheduled audit takes place. This can be 1 through 31 or LAST. This field is required if the frequency parameter is set to MONTHLY. If days 29-31 are specified, and the month does not have that many days, the audit takes place on the "LAST" day of the month.

$sel:targetCheckNames:UpdateScheduledAudit', updateScheduledAudit_targetCheckNames - Which checks are performed during the scheduled audit. Checks must be enabled for your account. (Use DescribeAccountAuditConfiguration to see the list of all checks, including those that are enabled or use UpdateAccountAuditConfiguration to select which checks are enabled.)

$sel:dayOfWeek:UpdateScheduledAudit', updateScheduledAudit_dayOfWeek - The day of the week on which the scheduled audit takes place. This can be one of SUN, MON, TUE, WED, THU, FRI, or SAT. This field is required if the "frequency" parameter is set to WEEKLY or BIWEEKLY.

$sel:scheduledAuditName:UpdateScheduledAudit', updateScheduledAudit_scheduledAuditName - The name of the scheduled audit. (Max. 128 chars)

Request Lenses

updateScheduledAudit_frequency :: Lens' UpdateScheduledAudit (Maybe AuditFrequency)

How often the scheduled audit takes place, either DAILY, WEEKLY, BIWEEKLY, or MONTHLY. The start time of each audit is determined by the system.

updateScheduledAudit_dayOfMonth :: Lens' UpdateScheduledAudit (Maybe Text)

The day of the month on which the scheduled audit takes place. This can be 1 through 31 or LAST. This field is required if the frequency parameter is set to MONTHLY. If days 29-31 are specified, and the month does not have that many days, the audit takes place on the "LAST" day of the month.

updateScheduledAudit_targetCheckNames :: Lens' UpdateScheduledAudit (Maybe [Text])

Which checks are performed during the scheduled audit. Checks must be enabled for your account. (Use DescribeAccountAuditConfiguration to see the list of all checks, including those that are enabled or use UpdateAccountAuditConfiguration to select which checks are enabled.)

updateScheduledAudit_dayOfWeek :: Lens' UpdateScheduledAudit (Maybe DayOfWeek)

The day of the week on which the scheduled audit takes place. This can be one of SUN, MON, TUE, WED, THU, FRI, or SAT. This field is required if the "frequency" parameter is set to WEEKLY or BIWEEKLY.

updateScheduledAudit_scheduledAuditName :: Lens' UpdateScheduledAudit Text

The name of the scheduled audit. (Max. 128 chars)

Destructuring the Response

data UpdateScheduledAuditResponse

See: newUpdateScheduledAuditResponse smart constructor.

Constructors

UpdateScheduledAuditResponse'

Fields

scheduledAuditArn :: Maybe Text

The ARN of the scheduled audit.

httpStatus :: Int

The response's http status code.

newUpdateScheduledAuditResponse

Arguments

:: Int
$sel:httpStatus:UpdateScheduledAuditResponse'

-> UpdateScheduledAuditResponse

Create a value of UpdateScheduledAuditResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scheduledAuditArn:UpdateScheduledAuditResponse', updateScheduledAuditResponse_scheduledAuditArn - The ARN of the scheduled audit.

$sel:httpStatus:UpdateScheduledAuditResponse', updateScheduledAuditResponse_httpStatus - The response's http status code.

Response Lenses

updateScheduledAuditResponse_scheduledAuditArn :: Lens' UpdateScheduledAuditResponse (Maybe Text)

The ARN of the scheduled audit.

updateScheduledAuditResponse_httpStatus :: Lens' UpdateScheduledAuditResponse Int

The response's http status code.