Network.AWS.IoT.UpdateSecurityProfile

Description

Updates a Device Defender security profile.

Requires permission to access the UpdateSecurityProfile action.

Creating a Request

data UpdateSecurityProfile

See: newUpdateSecurityProfile smart constructor.

Constructors

UpdateSecurityProfile'

Fields

expectedVersion :: Maybe Integer

The expected version of the security profile. A new version is generated whenever the security profile is updated. If you specify a value that is different from the actual version, a VersionConflictException is thrown.

alertTargets :: Maybe (HashMap AlertTargetType AlertTarget)

Where the alerts are sent. (Alerts are always sent to the console.)

deleteAdditionalMetricsToRetain :: Maybe Bool

If true, delete all additionalMetricsToRetain defined for this security profile. If any additionalMetricsToRetain are defined in the current invocation, an exception occurs.

additionalMetricsToRetain :: Maybe [Text]

/Please use UpdateSecurityProfileRequest$additionalMetricsToRetainV2 instead./

A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the profile's behaviors, but it is also retained for any metric specified here. Can be used with custom metrics; cannot be used with dimensions.

deleteAlertTargets :: Maybe Bool

If true, delete all alertTargets defined for this security profile. If any alertTargets are defined in the current invocation, an exception occurs.

behaviors :: Maybe [Behavior]

Specifies the behaviors that, when violated by a device (thing), cause an alert.

additionalMetricsToRetainV2 :: Maybe [MetricToRetain]

A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the profile's behaviors, but it is also retained for any metric specified here. Can be used with custom metrics; cannot be used with dimensions.

deleteBehaviors :: Maybe Bool

If true, delete all behaviors defined for this security profile. If any behaviors are defined in the current invocation, an exception occurs.

securityProfileDescription :: Maybe Text

A description of the security profile.

securityProfileName :: Text

The name of the security profile you want to update.

newUpdateSecurityProfile

Arguments

:: Text
$sel:securityProfileName:UpdateSecurityProfile'

-> UpdateSecurityProfile

Create a value of UpdateSecurityProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedVersion:UpdateSecurityProfile', updateSecurityProfile_expectedVersion - The expected version of the security profile. A new version is generated whenever the security profile is updated. If you specify a value that is different from the actual version, a VersionConflictException is thrown.

$sel:alertTargets:UpdateSecurityProfile', updateSecurityProfile_alertTargets - Where the alerts are sent. (Alerts are always sent to the console.)

$sel:deleteAdditionalMetricsToRetain:UpdateSecurityProfile', updateSecurityProfile_deleteAdditionalMetricsToRetain - If true, delete all additionalMetricsToRetain defined for this security profile. If any additionalMetricsToRetain are defined in the current invocation, an exception occurs.

$sel:additionalMetricsToRetain:UpdateSecurityProfile', updateSecurityProfile_additionalMetricsToRetain - /Please use UpdateSecurityProfileRequest$additionalMetricsToRetainV2 instead./

A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the profile's behaviors, but it is also retained for any metric specified here. Can be used with custom metrics; cannot be used with dimensions.

$sel:deleteAlertTargets:UpdateSecurityProfile', updateSecurityProfile_deleteAlertTargets - If true, delete all alertTargets defined for this security profile. If any alertTargets are defined in the current invocation, an exception occurs.

$sel:behaviors:UpdateSecurityProfile', updateSecurityProfile_behaviors - Specifies the behaviors that, when violated by a device (thing), cause an alert.

$sel:additionalMetricsToRetainV2:UpdateSecurityProfile', updateSecurityProfile_additionalMetricsToRetainV2 - A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the profile's behaviors, but it is also retained for any metric specified here. Can be used with custom metrics; cannot be used with dimensions.

$sel:deleteBehaviors:UpdateSecurityProfile', updateSecurityProfile_deleteBehaviors - If true, delete all behaviors defined for this security profile. If any behaviors are defined in the current invocation, an exception occurs.

$sel:securityProfileDescription:UpdateSecurityProfile', updateSecurityProfile_securityProfileDescription - A description of the security profile.

$sel:securityProfileName:UpdateSecurityProfile', updateSecurityProfile_securityProfileName - The name of the security profile you want to update.

Request Lenses

updateSecurityProfile_expectedVersion :: Lens' UpdateSecurityProfile (Maybe Integer)

The expected version of the security profile. A new version is generated whenever the security profile is updated. If you specify a value that is different from the actual version, a VersionConflictException is thrown.

updateSecurityProfile_alertTargets :: Lens' UpdateSecurityProfile (Maybe (HashMap AlertTargetType AlertTarget))

Where the alerts are sent. (Alerts are always sent to the console.)

updateSecurityProfile_deleteAdditionalMetricsToRetain :: Lens' UpdateSecurityProfile (Maybe Bool)

If true, delete all additionalMetricsToRetain defined for this security profile. If any additionalMetricsToRetain are defined in the current invocation, an exception occurs.

updateSecurityProfile_additionalMetricsToRetain :: Lens' UpdateSecurityProfile (Maybe [Text])

/Please use UpdateSecurityProfileRequest$additionalMetricsToRetainV2 instead./

A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the profile's behaviors, but it is also retained for any metric specified here. Can be used with custom metrics; cannot be used with dimensions.

updateSecurityProfile_deleteAlertTargets :: Lens' UpdateSecurityProfile (Maybe Bool)

If true, delete all alertTargets defined for this security profile. If any alertTargets are defined in the current invocation, an exception occurs.

updateSecurityProfile_behaviors :: Lens' UpdateSecurityProfile (Maybe [Behavior])

Specifies the behaviors that, when violated by a device (thing), cause an alert.

updateSecurityProfile_additionalMetricsToRetainV2 :: Lens' UpdateSecurityProfile (Maybe [MetricToRetain])

A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the profile's behaviors, but it is also retained for any metric specified here. Can be used with custom metrics; cannot be used with dimensions.

updateSecurityProfile_deleteBehaviors :: Lens' UpdateSecurityProfile (Maybe Bool)

If true, delete all behaviors defined for this security profile. If any behaviors are defined in the current invocation, an exception occurs.

updateSecurityProfile_securityProfileDescription :: Lens' UpdateSecurityProfile (Maybe Text)

A description of the security profile.

updateSecurityProfile_securityProfileName :: Lens' UpdateSecurityProfile Text

The name of the security profile you want to update.

Destructuring the Response

data UpdateSecurityProfileResponse

See: newUpdateSecurityProfileResponse smart constructor.

Constructors

UpdateSecurityProfileResponse'

Fields

lastModifiedDate :: Maybe POSIX

The time the security profile was last modified.

alertTargets :: Maybe (HashMap AlertTargetType AlertTarget)

Where the alerts are sent. (Alerts are always sent to the console.)

additionalMetricsToRetain :: Maybe [Text]

/Please use UpdateSecurityProfileResponse$additionalMetricsToRetainV2 instead./

A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the security profile's behaviors, but it is also retained for any metric specified here.

creationDate :: Maybe POSIX

The time the security profile was created.

securityProfileName :: Maybe Text

The name of the security profile that was updated.

version :: Maybe Integer

The updated version of the security profile.

behaviors :: Maybe [Behavior]

Specifies the behaviors that, when violated by a device (thing), cause an alert.

additionalMetricsToRetainV2 :: Maybe [MetricToRetain]

A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the profile's behaviors, but it is also retained for any metric specified here. Can be used with custom metrics; cannot be used with dimensions.

securityProfileDescription :: Maybe Text

The description of the security profile.

securityProfileArn :: Maybe Text

The ARN of the security profile that was updated.

httpStatus :: Int

The response's http status code.

newUpdateSecurityProfileResponse

Arguments

:: Int
$sel:httpStatus:UpdateSecurityProfileResponse'

-> UpdateSecurityProfileResponse

Create a value of UpdateSecurityProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:UpdateSecurityProfileResponse', updateSecurityProfileResponse_lastModifiedDate - The time the security profile was last modified.

$sel:alertTargets:UpdateSecurityProfile', updateSecurityProfileResponse_alertTargets - Where the alerts are sent. (Alerts are always sent to the console.)

$sel:additionalMetricsToRetain:UpdateSecurityProfile', updateSecurityProfileResponse_additionalMetricsToRetain - /Please use UpdateSecurityProfileResponse$additionalMetricsToRetainV2 instead./

A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the security profile's behaviors, but it is also retained for any metric specified here.

$sel:creationDate:UpdateSecurityProfileResponse', updateSecurityProfileResponse_creationDate - The time the security profile was created.

$sel:securityProfileName:UpdateSecurityProfile', updateSecurityProfileResponse_securityProfileName - The name of the security profile that was updated.

$sel:version:UpdateSecurityProfileResponse', updateSecurityProfileResponse_version - The updated version of the security profile.

$sel:behaviors:UpdateSecurityProfile', updateSecurityProfileResponse_behaviors - Specifies the behaviors that, when violated by a device (thing), cause an alert.

$sel:additionalMetricsToRetainV2:UpdateSecurityProfile', updateSecurityProfileResponse_additionalMetricsToRetainV2 - A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the profile's behaviors, but it is also retained for any metric specified here. Can be used with custom metrics; cannot be used with dimensions.

$sel:securityProfileDescription:UpdateSecurityProfile', updateSecurityProfileResponse_securityProfileDescription - The description of the security profile.

$sel:securityProfileArn:UpdateSecurityProfileResponse', updateSecurityProfileResponse_securityProfileArn - The ARN of the security profile that was updated.

$sel:httpStatus:UpdateSecurityProfileResponse', updateSecurityProfileResponse_httpStatus - The response's http status code.

Response Lenses

updateSecurityProfileResponse_lastModifiedDate :: Lens' UpdateSecurityProfileResponse (Maybe UTCTime)

The time the security profile was last modified.

updateSecurityProfileResponse_alertTargets :: Lens' UpdateSecurityProfileResponse (Maybe (HashMap AlertTargetType AlertTarget))

Where the alerts are sent. (Alerts are always sent to the console.)

updateSecurityProfileResponse_additionalMetricsToRetain :: Lens' UpdateSecurityProfileResponse (Maybe [Text])

/Please use UpdateSecurityProfileResponse$additionalMetricsToRetainV2 instead./

A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the security profile's behaviors, but it is also retained for any metric specified here.

updateSecurityProfileResponse_creationDate :: Lens' UpdateSecurityProfileResponse (Maybe UTCTime)

The time the security profile was created.

updateSecurityProfileResponse_securityProfileName :: Lens' UpdateSecurityProfileResponse (Maybe Text)

The name of the security profile that was updated.

updateSecurityProfileResponse_version :: Lens' UpdateSecurityProfileResponse (Maybe Integer)

The updated version of the security profile.

updateSecurityProfileResponse_behaviors :: Lens' UpdateSecurityProfileResponse (Maybe [Behavior])

Specifies the behaviors that, when violated by a device (thing), cause an alert.

updateSecurityProfileResponse_additionalMetricsToRetainV2 :: Lens' UpdateSecurityProfileResponse (Maybe [MetricToRetain])

A list of metrics whose data is retained (stored). By default, data is retained for any metric used in the profile's behaviors, but it is also retained for any metric specified here. Can be used with custom metrics; cannot be used with dimensions.

updateSecurityProfileResponse_securityProfileDescription :: Lens' UpdateSecurityProfileResponse (Maybe Text)

The description of the security profile.

updateSecurityProfileResponse_securityProfileArn :: Lens' UpdateSecurityProfileResponse (Maybe Text)

The ARN of the security profile that was updated.

updateSecurityProfileResponse_httpStatus :: Lens' UpdateSecurityProfileResponse Int

The response's http status code.