Network.AWS.IoT.UpdateThing

Description

Updates the data for a thing.

Requires permission to access the UpdateThing action.

Creating a Request

data UpdateThing

The input for the UpdateThing operation.

See: newUpdateThing smart constructor.

Constructors

UpdateThing'

Fields

expectedVersion :: Maybe Integer

The expected version of the thing record in the registry. If the version of the record in the registry does not match the expected version specified in the request, the UpdateThing request is rejected with a VersionConflictException.

thingTypeName :: Maybe Text

The name of the thing type.

attributePayload :: Maybe AttributePayload

A list of thing attributes, a JSON string containing name-value pairs. For example:

{\"attributes\":{\"name1\":\"value2\"}}

This data is used to add new attributes or update existing attributes.

removeThingType :: Maybe Bool

Remove a thing type association. If true, the association is removed.

thingName :: Text

The name of the thing to update.

You can't change a thing's name. To change a thing's name, you must create a new thing, give it the new name, and then delete the old thing.

newUpdateThing

Arguments

:: Text
$sel:thingName:UpdateThing'

-> UpdateThing

Create a value of UpdateThing with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expectedVersion:UpdateThing', updateThing_expectedVersion - The expected version of the thing record in the registry. If the version of the record in the registry does not match the expected version specified in the request, the UpdateThing request is rejected with a VersionConflictException.

$sel:thingTypeName:UpdateThing', updateThing_thingTypeName - The name of the thing type.

$sel:attributePayload:UpdateThing', updateThing_attributePayload - A list of thing attributes, a JSON string containing name-value pairs. For example:

{\"attributes\":{\"name1\":\"value2\"}}

This data is used to add new attributes or update existing attributes.

$sel:removeThingType:UpdateThing', updateThing_removeThingType - Remove a thing type association. If true, the association is removed.

$sel:thingName:UpdateThing', updateThing_thingName - The name of the thing to update.

You can't change a thing's name. To change a thing's name, you must create a new thing, give it the new name, and then delete the old thing.

Request Lenses

updateThing_expectedVersion :: Lens' UpdateThing (Maybe Integer)

The expected version of the thing record in the registry. If the version of the record in the registry does not match the expected version specified in the request, the UpdateThing request is rejected with a VersionConflictException.

updateThing_thingTypeName :: Lens' UpdateThing (Maybe Text)

The name of the thing type.

updateThing_attributePayload :: Lens' UpdateThing (Maybe AttributePayload)

A list of thing attributes, a JSON string containing name-value pairs. For example:

{\"attributes\":{\"name1\":\"value2\"}}

This data is used to add new attributes or update existing attributes.

updateThing_removeThingType :: Lens' UpdateThing (Maybe Bool)

Remove a thing type association. If true, the association is removed.

updateThing_thingName :: Lens' UpdateThing Text

The name of the thing to update.

You can't change a thing's name. To change a thing's name, you must create a new thing, give it the new name, and then delete the old thing.

Destructuring the Response

data UpdateThingResponse

The output from the UpdateThing operation.

See: newUpdateThingResponse smart constructor.

Constructors

UpdateThingResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateThingResponse

Arguments

:: Int
$sel:httpStatus:UpdateThingResponse'

-> UpdateThingResponse

Create a value of UpdateThingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateThingResponse', updateThingResponse_httpStatus - The response's http status code.

Response Lenses

updateThingResponse_httpStatus :: Lens' UpdateThingResponse Int

The response's http status code.