Network.AWS.IoT.UpdateThingGroupsForThing

Description

Updates the groups to which the thing belongs.

Requires permission to access the UpdateThingGroupsForThing action.

Creating a Request

data UpdateThingGroupsForThing

See: newUpdateThingGroupsForThing smart constructor.

Constructors

UpdateThingGroupsForThing'

Fields

thingGroupsToAdd :: Maybe [Text]

The groups to which the thing will be added.

thingGroupsToRemove :: Maybe [Text]

The groups from which the thing will be removed.

overrideDynamicGroups :: Maybe Bool

Override dynamic thing groups with static thing groups when 10-group limit is reached. If a thing belongs to 10 thing groups, and one or more of those groups are dynamic thing groups, adding a thing to a static group removes the thing from the last dynamic group.

thingName :: Maybe Text

The thing whose group memberships will be updated.

newUpdateThingGroupsForThing :: UpdateThingGroupsForThing

Create a value of UpdateThingGroupsForThing with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thingGroupsToAdd:UpdateThingGroupsForThing', updateThingGroupsForThing_thingGroupsToAdd - The groups to which the thing will be added.

$sel:thingGroupsToRemove:UpdateThingGroupsForThing', updateThingGroupsForThing_thingGroupsToRemove - The groups from which the thing will be removed.

$sel:overrideDynamicGroups:UpdateThingGroupsForThing', updateThingGroupsForThing_overrideDynamicGroups - Override dynamic thing groups with static thing groups when 10-group limit is reached. If a thing belongs to 10 thing groups, and one or more of those groups are dynamic thing groups, adding a thing to a static group removes the thing from the last dynamic group.

$sel:thingName:UpdateThingGroupsForThing', updateThingGroupsForThing_thingName - The thing whose group memberships will be updated.

Request Lenses

updateThingGroupsForThing_thingGroupsToAdd :: Lens' UpdateThingGroupsForThing (Maybe [Text])

The groups to which the thing will be added.

updateThingGroupsForThing_thingGroupsToRemove :: Lens' UpdateThingGroupsForThing (Maybe [Text])

The groups from which the thing will be removed.

updateThingGroupsForThing_overrideDynamicGroups :: Lens' UpdateThingGroupsForThing (Maybe Bool)

Override dynamic thing groups with static thing groups when 10-group limit is reached. If a thing belongs to 10 thing groups, and one or more of those groups are dynamic thing groups, adding a thing to a static group removes the thing from the last dynamic group.

updateThingGroupsForThing_thingName :: Lens' UpdateThingGroupsForThing (Maybe Text)

The thing whose group memberships will be updated.

Destructuring the Response

data UpdateThingGroupsForThingResponse

See: newUpdateThingGroupsForThingResponse smart constructor.

Constructors

UpdateThingGroupsForThingResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateThingGroupsForThingResponse

Arguments

:: Int
$sel:httpStatus:UpdateThingGroupsForThingResponse'

-> UpdateThingGroupsForThingResponse

Create a value of UpdateThingGroupsForThingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateThingGroupsForThingResponse', updateThingGroupsForThingResponse_httpStatus - The response's http status code.

Response Lenses

updateThingGroupsForThingResponse_httpStatus :: Lens' UpdateThingGroupsForThingResponse Int

The response's http status code.