Network.AWS.IoT.UpdateTopicRuleDestination

Description

Updates a topic rule destination. You use this to change the status, endpoint URL, or confirmation URL of the destination.

Requires permission to access the UpdateTopicRuleDestination action.

Creating a Request

data UpdateTopicRuleDestination

See: newUpdateTopicRuleDestination smart constructor.

Constructors

UpdateTopicRuleDestination'

Fields

arn :: Text

The ARN of the topic rule destination.

status :: TopicRuleDestinationStatus

The status of the topic rule destination. Valid values are:

IN_PROGRESS

A topic rule destination was created but has not been confirmed. You can set status to IN_PROGRESS by calling UpdateTopicRuleDestination. Calling UpdateTopicRuleDestination causes a new confirmation challenge to be sent to your confirmation endpoint.

ENABLED

Confirmation was completed, and traffic to this destination is allowed. You can set status to DISABLED by calling UpdateTopicRuleDestination.

DISABLED

Confirmation was completed, and traffic to this destination is not allowed. You can set status to ENABLED by calling UpdateTopicRuleDestination.

ERROR

Confirmation could not be completed, for example if the confirmation timed out. You can call GetTopicRuleDestination for details about the error. You can set status to IN_PROGRESS by calling UpdateTopicRuleDestination. Calling UpdateTopicRuleDestination causes a new confirmation challenge to be sent to your confirmation endpoint.

newUpdateTopicRuleDestination

Arguments

:: Text
$sel:arn:UpdateTopicRuleDestination'

-> TopicRuleDestinationStatus
$sel:status:UpdateTopicRuleDestination'

-> UpdateTopicRuleDestination

Create a value of UpdateTopicRuleDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:UpdateTopicRuleDestination', updateTopicRuleDestination_arn - The ARN of the topic rule destination.

$sel:status:UpdateTopicRuleDestination', updateTopicRuleDestination_status - The status of the topic rule destination. Valid values are:

IN_PROGRESS

A topic rule destination was created but has not been confirmed. You can set status to IN_PROGRESS by calling UpdateTopicRuleDestination. Calling UpdateTopicRuleDestination causes a new confirmation challenge to be sent to your confirmation endpoint.

ENABLED

Confirmation was completed, and traffic to this destination is allowed. You can set status to DISABLED by calling UpdateTopicRuleDestination.

DISABLED

Confirmation was completed, and traffic to this destination is not allowed. You can set status to ENABLED by calling UpdateTopicRuleDestination.

ERROR

Confirmation could not be completed, for example if the confirmation timed out. You can call GetTopicRuleDestination for details about the error. You can set status to IN_PROGRESS by calling UpdateTopicRuleDestination. Calling UpdateTopicRuleDestination causes a new confirmation challenge to be sent to your confirmation endpoint.

Request Lenses

updateTopicRuleDestination_arn :: Lens' UpdateTopicRuleDestination Text

The ARN of the topic rule destination.

updateTopicRuleDestination_status :: Lens' UpdateTopicRuleDestination TopicRuleDestinationStatus

The status of the topic rule destination. Valid values are:

IN_PROGRESS

A topic rule destination was created but has not been confirmed. You can set status to IN_PROGRESS by calling UpdateTopicRuleDestination. Calling UpdateTopicRuleDestination causes a new confirmation challenge to be sent to your confirmation endpoint.

ENABLED

Confirmation was completed, and traffic to this destination is allowed. You can set status to DISABLED by calling UpdateTopicRuleDestination.

DISABLED

Confirmation was completed, and traffic to this destination is not allowed. You can set status to ENABLED by calling UpdateTopicRuleDestination.

ERROR

Confirmation could not be completed, for example if the confirmation timed out. You can call GetTopicRuleDestination for details about the error. You can set status to IN_PROGRESS by calling UpdateTopicRuleDestination. Calling UpdateTopicRuleDestination causes a new confirmation challenge to be sent to your confirmation endpoint.

Destructuring the Response

data UpdateTopicRuleDestinationResponse

See: newUpdateTopicRuleDestinationResponse smart constructor.

Constructors

UpdateTopicRuleDestinationResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateTopicRuleDestinationResponse

Arguments

:: Int
$sel:httpStatus:UpdateTopicRuleDestinationResponse'

-> UpdateTopicRuleDestinationResponse

Create a value of UpdateTopicRuleDestinationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateTopicRuleDestinationResponse', updateTopicRuleDestinationResponse_httpStatus - The response's http status code.

Response Lenses

updateTopicRuleDestinationResponse_httpStatus :: Lens' UpdateTopicRuleDestinationResponse Int

The response's http status code.