Network.AWS.IoT.ValidateSecurityProfileBehaviors

Description

Validates a Device Defender security profile behaviors specification.

Requires permission to access the ValidateSecurityProfileBehaviors action.

Creating a Request

data ValidateSecurityProfileBehaviors

See: newValidateSecurityProfileBehaviors smart constructor.

Constructors

ValidateSecurityProfileBehaviors'

Fields

behaviors :: [Behavior]

Specifies the behaviors that, when violated by a device (thing), cause an alert.

newValidateSecurityProfileBehaviors :: ValidateSecurityProfileBehaviors

Create a value of ValidateSecurityProfileBehaviors with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:behaviors:ValidateSecurityProfileBehaviors', validateSecurityProfileBehaviors_behaviors - Specifies the behaviors that, when violated by a device (thing), cause an alert.

Request Lenses

validateSecurityProfileBehaviors_behaviors :: Lens' ValidateSecurityProfileBehaviors [Behavior]

Specifies the behaviors that, when violated by a device (thing), cause an alert.

Destructuring the Response

data ValidateSecurityProfileBehaviorsResponse

See: newValidateSecurityProfileBehaviorsResponse smart constructor.

Constructors

ValidateSecurityProfileBehaviorsResponse'

Fields

validationErrors :: Maybe [ValidationError]

The list of any errors found in the behaviors.

valid :: Maybe Bool

True if the behaviors were valid.

httpStatus :: Int

The response's http status code.

newValidateSecurityProfileBehaviorsResponse

Arguments

:: Int
$sel:httpStatus:ValidateSecurityProfileBehaviorsResponse'

-> ValidateSecurityProfileBehaviorsResponse

Create a value of ValidateSecurityProfileBehaviorsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:validationErrors:ValidateSecurityProfileBehaviorsResponse', validateSecurityProfileBehaviorsResponse_validationErrors - The list of any errors found in the behaviors.

$sel:valid:ValidateSecurityProfileBehaviorsResponse', validateSecurityProfileBehaviorsResponse_valid - True if the behaviors were valid.

$sel:httpStatus:ValidateSecurityProfileBehaviorsResponse', validateSecurityProfileBehaviorsResponse_httpStatus - The response's http status code.

Response Lenses

validateSecurityProfileBehaviorsResponse_validationErrors :: Lens' ValidateSecurityProfileBehaviorsResponse (Maybe [ValidationError])

The list of any errors found in the behaviors.

validateSecurityProfileBehaviorsResponse_valid :: Lens' ValidateSecurityProfileBehaviorsResponse (Maybe Bool)

True if the behaviors were valid.

validateSecurityProfileBehaviorsResponse_httpStatus :: Lens' ValidateSecurityProfileBehaviorsResponse Int

The response's http status code.