Network.AWS.IoT1ClickDevices.Types.DeviceMethod

Description

Documentation

data DeviceMethod

See: newDeviceMethod smart constructor.

Constructors

DeviceMethod'

Fields

methodName :: Maybe Text

The name of the method applicable to the deviceType.

deviceType :: Maybe Text

The type of the device, such as "button".

newDeviceMethod :: DeviceMethod

Create a value of DeviceMethod with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:methodName:DeviceMethod', deviceMethod_methodName - The name of the method applicable to the deviceType.

$sel:deviceType:DeviceMethod', deviceMethod_deviceType - The type of the device, such as "button".

deviceMethod_methodName :: Lens' DeviceMethod (Maybe Text)

The name of the method applicable to the deviceType.

deviceMethod_deviceType :: Lens' DeviceMethod (Maybe Text)

The type of the device, such as "button".