Network.AWS.IoT1ClickProjects.AssociateDeviceWithPlacement

Description

Associates a physical device with a placement.

Creating a Request

data AssociateDeviceWithPlacement

See: newAssociateDeviceWithPlacement smart constructor.

Constructors

AssociateDeviceWithPlacement'

Fields

projectName :: Text

The name of the project containing the placement in which to associate the device.

placementName :: Text

The name of the placement in which to associate the device.

deviceId :: Text

The ID of the physical device to be associated with the given placement in the project. Note that a mandatory 4 character prefix is required for all deviceId values.

deviceTemplateName :: Text

The device template name to associate with the device ID.

newAssociateDeviceWithPlacement

Arguments

:: Text
$sel:projectName:AssociateDeviceWithPlacement'

-> Text
$sel:placementName:AssociateDeviceWithPlacement'

-> Text
$sel:deviceId:AssociateDeviceWithPlacement'

-> Text
$sel:deviceTemplateName:AssociateDeviceWithPlacement'

-> AssociateDeviceWithPlacement

Create a value of AssociateDeviceWithPlacement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectName:AssociateDeviceWithPlacement', associateDeviceWithPlacement_projectName - The name of the project containing the placement in which to associate the device.

$sel:placementName:AssociateDeviceWithPlacement', associateDeviceWithPlacement_placementName - The name of the placement in which to associate the device.

$sel:deviceId:AssociateDeviceWithPlacement', associateDeviceWithPlacement_deviceId - The ID of the physical device to be associated with the given placement in the project. Note that a mandatory 4 character prefix is required for all deviceId values.

$sel:deviceTemplateName:AssociateDeviceWithPlacement', associateDeviceWithPlacement_deviceTemplateName - The device template name to associate with the device ID.

Request Lenses

associateDeviceWithPlacement_projectName :: Lens' AssociateDeviceWithPlacement Text

The name of the project containing the placement in which to associate the device.

associateDeviceWithPlacement_placementName :: Lens' AssociateDeviceWithPlacement Text

The name of the placement in which to associate the device.

associateDeviceWithPlacement_deviceId :: Lens' AssociateDeviceWithPlacement Text

The ID of the physical device to be associated with the given placement in the project. Note that a mandatory 4 character prefix is required for all deviceId values.

associateDeviceWithPlacement_deviceTemplateName :: Lens' AssociateDeviceWithPlacement Text

The device template name to associate with the device ID.

Destructuring the Response

data AssociateDeviceWithPlacementResponse

See: newAssociateDeviceWithPlacementResponse smart constructor.

Constructors

AssociateDeviceWithPlacementResponse'

Fields

httpStatus :: Int

The response's http status code.

newAssociateDeviceWithPlacementResponse

Arguments

:: Int
$sel:httpStatus:AssociateDeviceWithPlacementResponse'

-> AssociateDeviceWithPlacementResponse

Create a value of AssociateDeviceWithPlacementResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AssociateDeviceWithPlacementResponse', associateDeviceWithPlacementResponse_httpStatus - The response's http status code.

Response Lenses

associateDeviceWithPlacementResponse_httpStatus :: Lens' AssociateDeviceWithPlacementResponse Int

The response's http status code.