Network.AWS.IoT1ClickProjects.DisassociateDeviceFromPlacement

Description

Removes a physical device from a placement.

Creating a Request

data DisassociateDeviceFromPlacement

See: newDisassociateDeviceFromPlacement smart constructor.

Constructors

DisassociateDeviceFromPlacement'

Fields

projectName :: Text

The name of the project that contains the placement.

placementName :: Text

The name of the placement that the device should be removed from.

deviceTemplateName :: Text

The device ID that should be removed from the placement.

newDisassociateDeviceFromPlacement

Arguments

:: Text
$sel:projectName:DisassociateDeviceFromPlacement'

-> Text
$sel:placementName:DisassociateDeviceFromPlacement'

-> Text
$sel:deviceTemplateName:DisassociateDeviceFromPlacement'

-> DisassociateDeviceFromPlacement

Create a value of DisassociateDeviceFromPlacement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectName:DisassociateDeviceFromPlacement', disassociateDeviceFromPlacement_projectName - The name of the project that contains the placement.

$sel:placementName:DisassociateDeviceFromPlacement', disassociateDeviceFromPlacement_placementName - The name of the placement that the device should be removed from.

$sel:deviceTemplateName:DisassociateDeviceFromPlacement', disassociateDeviceFromPlacement_deviceTemplateName - The device ID that should be removed from the placement.

Request Lenses

disassociateDeviceFromPlacement_projectName :: Lens' DisassociateDeviceFromPlacement Text

The name of the project that contains the placement.

disassociateDeviceFromPlacement_placementName :: Lens' DisassociateDeviceFromPlacement Text

The name of the placement that the device should be removed from.

disassociateDeviceFromPlacement_deviceTemplateName :: Lens' DisassociateDeviceFromPlacement Text

The device ID that should be removed from the placement.

Destructuring the Response

data DisassociateDeviceFromPlacementResponse

See: newDisassociateDeviceFromPlacementResponse smart constructor.

Constructors

DisassociateDeviceFromPlacementResponse'

Fields

httpStatus :: Int

The response's http status code.

newDisassociateDeviceFromPlacementResponse

Arguments

:: Int
$sel:httpStatus:DisassociateDeviceFromPlacementResponse'

-> DisassociateDeviceFromPlacementResponse

Create a value of DisassociateDeviceFromPlacementResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisassociateDeviceFromPlacementResponse', disassociateDeviceFromPlacementResponse_httpStatus - The response's http status code.

Response Lenses

disassociateDeviceFromPlacementResponse_httpStatus :: Lens' DisassociateDeviceFromPlacementResponse Int

The response's http status code.