Network.AWS.IoT1ClickProjects.ListPlacements

Description

Lists the placement(s) of a project.

This operation returns paginated results.

Creating a Request

data ListPlacements

See: newListPlacements smart constructor.

Constructors

ListPlacements'

Fields

nextToken :: Maybe Text

The token to retrieve the next set of results.

maxResults :: Maybe Natural

The maximum number of results to return per request. If not set, a default value of 100 is used.

projectName :: Text

The project containing the placements to be listed.

newListPlacements

Arguments

:: Text
$sel:projectName:ListPlacements'

-> ListPlacements

Create a value of ListPlacements with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPlacements', listPlacements_nextToken - The token to retrieve the next set of results.

$sel:maxResults:ListPlacements', listPlacements_maxResults - The maximum number of results to return per request. If not set, a default value of 100 is used.

$sel:projectName:ListPlacements', listPlacements_projectName - The project containing the placements to be listed.

Request Lenses

listPlacements_nextToken :: Lens' ListPlacements (Maybe Text)

The token to retrieve the next set of results.

listPlacements_maxResults :: Lens' ListPlacements (Maybe Natural)

The maximum number of results to return per request. If not set, a default value of 100 is used.

listPlacements_projectName :: Lens' ListPlacements Text

The project containing the placements to be listed.

Destructuring the Response

data ListPlacementsResponse

See: newListPlacementsResponse smart constructor.

Constructors

ListPlacementsResponse'

Fields

nextToken :: Maybe Text

The token used to retrieve the next set of results - will be effectively empty if there are no further results.

httpStatus :: Int

The response's http status code.

placements :: [PlacementSummary]

An object listing the requested placements.

newListPlacementsResponse

Arguments

:: Int
$sel:httpStatus:ListPlacementsResponse'

-> ListPlacementsResponse

Create a value of ListPlacementsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPlacements', listPlacementsResponse_nextToken - The token used to retrieve the next set of results - will be effectively empty if there are no further results.

$sel:httpStatus:ListPlacementsResponse', listPlacementsResponse_httpStatus - The response's http status code.

$sel:placements:ListPlacementsResponse', listPlacementsResponse_placements - An object listing the requested placements.

Response Lenses

listPlacementsResponse_nextToken :: Lens' ListPlacementsResponse (Maybe Text)

The token used to retrieve the next set of results - will be effectively empty if there are no further results.

listPlacementsResponse_httpStatus :: Lens' ListPlacementsResponse Int

The response's http status code.

listPlacementsResponse_placements :: Lens' ListPlacementsResponse [PlacementSummary]

An object listing the requested placements.