Network.AWS.IoT1ClickProjects.Types

Description

Service Configuration

defaultService :: Service

API version 2018-05-14 of the Amazon IoT 1-Click Projects Service SDK configuration.

Errors

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

_InternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

_ResourceConflictException :: AsError a => Getting (First ServiceError) a ServiceError

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

DeviceTemplate

data DeviceTemplate

An object representing a device for a placement template (see PlacementTemplate).

See: newDeviceTemplate smart constructor.

Constructors

DeviceTemplate'

Fields

deviceType :: Maybe Text

The device type, which currently must be "button".

callbackOverrides :: Maybe (HashMap Text Text)

An optional Lambda function to invoke instead of the default Lambda function provided by the placement template.

newDeviceTemplate :: DeviceTemplate

Create a value of DeviceTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceType:DeviceTemplate', deviceTemplate_deviceType - The device type, which currently must be "button".

$sel:callbackOverrides:DeviceTemplate', deviceTemplate_callbackOverrides - An optional Lambda function to invoke instead of the default Lambda function provided by the placement template.

deviceTemplate_deviceType :: Lens' DeviceTemplate (Maybe Text)

The device type, which currently must be "button".

deviceTemplate_callbackOverrides :: Lens' DeviceTemplate (Maybe (HashMap Text Text))

An optional Lambda function to invoke instead of the default Lambda function provided by the placement template.

PlacementDescription

data PlacementDescription

An object describing a project's placement.

See: newPlacementDescription smart constructor.

Constructors

PlacementDescription'

Fields

projectName :: Text

The name of the project containing the placement.

placementName :: Text

The name of the placement.

attributes :: HashMap Text Text

The user-defined attributes associated with the placement.

createdDate :: POSIX

The date when the placement was initially created, in UNIX epoch time format.

updatedDate :: POSIX

The date when the placement was last updated, in UNIX epoch time format. If the placement was not updated, then createdDate and updatedDate are the same.

newPlacementDescription

Arguments

:: Text
$sel:projectName:PlacementDescription'

-> Text
$sel:placementName:PlacementDescription'

-> UTCTime
$sel:createdDate:PlacementDescription'

-> UTCTime
$sel:updatedDate:PlacementDescription'

-> PlacementDescription

Create a value of PlacementDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectName:PlacementDescription', placementDescription_projectName - The name of the project containing the placement.

$sel:placementName:PlacementDescription', placementDescription_placementName - The name of the placement.

$sel:attributes:PlacementDescription', placementDescription_attributes - The user-defined attributes associated with the placement.

$sel:createdDate:PlacementDescription', placementDescription_createdDate - The date when the placement was initially created, in UNIX epoch time format.

$sel:updatedDate:PlacementDescription', placementDescription_updatedDate - The date when the placement was last updated, in UNIX epoch time format. If the placement was not updated, then createdDate and updatedDate are the same.

placementDescription_projectName :: Lens' PlacementDescription Text

The name of the project containing the placement.

placementDescription_placementName :: Lens' PlacementDescription Text

The name of the placement.

placementDescription_attributes :: Lens' PlacementDescription (HashMap Text Text)

The user-defined attributes associated with the placement.

placementDescription_createdDate :: Lens' PlacementDescription UTCTime

The date when the placement was initially created, in UNIX epoch time format.

placementDescription_updatedDate :: Lens' PlacementDescription UTCTime

The date when the placement was last updated, in UNIX epoch time format. If the placement was not updated, then createdDate and updatedDate are the same.

PlacementSummary

data PlacementSummary

An object providing summary information for a particular placement.

See: newPlacementSummary smart constructor.

Constructors

PlacementSummary'

Fields

projectName :: Text

The name of the project containing the placement.

placementName :: Text

The name of the placement being summarized.

createdDate :: POSIX

The date when the placement was originally created, in UNIX epoch time format.

updatedDate :: POSIX

The date when the placement was last updated, in UNIX epoch time format. If the placement was not updated, then createdDate and updatedDate are the same.

newPlacementSummary

Arguments

:: Text
$sel:projectName:PlacementSummary'

-> Text
$sel:placementName:PlacementSummary'

-> UTCTime
$sel:createdDate:PlacementSummary'

-> UTCTime
$sel:updatedDate:PlacementSummary'

-> PlacementSummary

Create a value of PlacementSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectName:PlacementSummary', placementSummary_projectName - The name of the project containing the placement.

$sel:placementName:PlacementSummary', placementSummary_placementName - The name of the placement being summarized.

$sel:createdDate:PlacementSummary', placementSummary_createdDate - The date when the placement was originally created, in UNIX epoch time format.

$sel:updatedDate:PlacementSummary', placementSummary_updatedDate - The date when the placement was last updated, in UNIX epoch time format. If the placement was not updated, then createdDate and updatedDate are the same.

placementSummary_projectName :: Lens' PlacementSummary Text

The name of the project containing the placement.

placementSummary_placementName :: Lens' PlacementSummary Text

The name of the placement being summarized.

placementSummary_createdDate :: Lens' PlacementSummary UTCTime

The date when the placement was originally created, in UNIX epoch time format.

placementSummary_updatedDate :: Lens' PlacementSummary UTCTime

The date when the placement was last updated, in UNIX epoch time format. If the placement was not updated, then createdDate and updatedDate are the same.

PlacementTemplate

data PlacementTemplate

An object defining the template for a placement.

See: newPlacementTemplate smart constructor.

Constructors

PlacementTemplate'

Fields

deviceTemplates :: Maybe (HashMap Text DeviceTemplate)

An object specifying the DeviceTemplate for all placements using this (PlacementTemplate) template.

defaultAttributes :: Maybe (HashMap Text Text)

The default attributes (key/value pairs) to be applied to all placements using this template.

newPlacementTemplate :: PlacementTemplate

Create a value of PlacementTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceTemplates:PlacementTemplate', placementTemplate_deviceTemplates - An object specifying the DeviceTemplate for all placements using this (PlacementTemplate) template.

$sel:defaultAttributes:PlacementTemplate', placementTemplate_defaultAttributes - The default attributes (key/value pairs) to be applied to all placements using this template.

placementTemplate_deviceTemplates :: Lens' PlacementTemplate (Maybe (HashMap Text DeviceTemplate))

An object specifying the DeviceTemplate for all placements using this (PlacementTemplate) template.

placementTemplate_defaultAttributes :: Lens' PlacementTemplate (Maybe (HashMap Text Text))

The default attributes (key/value pairs) to be applied to all placements using this template.

ProjectDescription

data ProjectDescription

An object providing detailed information for a particular project associated with an AWS account and region.

See: newProjectDescription smart constructor.

Constructors

ProjectDescription'

Fields

arn :: Maybe Text

The ARN of the project.

placementTemplate :: Maybe PlacementTemplate

An object describing the project's placement specifications.

description :: Maybe Text

The description of the project.

tags :: Maybe (HashMap Text Text)

The tags (metadata key/value pairs) associated with the project.

projectName :: Text

The name of the project for which to obtain information from.

createdDate :: POSIX

The date when the project was originally created, in UNIX epoch time format.

updatedDate :: POSIX

The date when the project was last updated, in UNIX epoch time format. If the project was not updated, then createdDate and updatedDate are the same.

newProjectDescription

Arguments

:: Text
$sel:projectName:ProjectDescription'

-> UTCTime
$sel:createdDate:ProjectDescription'

-> UTCTime
$sel:updatedDate:ProjectDescription'

-> ProjectDescription

Create a value of ProjectDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:ProjectDescription', projectDescription_arn - The ARN of the project.

$sel:placementTemplate:ProjectDescription', projectDescription_placementTemplate - An object describing the project's placement specifications.

$sel:description:ProjectDescription', projectDescription_description - The description of the project.

$sel:tags:ProjectDescription', projectDescription_tags - The tags (metadata key/value pairs) associated with the project.

$sel:projectName:ProjectDescription', projectDescription_projectName - The name of the project for which to obtain information from.

$sel:createdDate:ProjectDescription', projectDescription_createdDate - The date when the project was originally created, in UNIX epoch time format.

$sel:updatedDate:ProjectDescription', projectDescription_updatedDate - The date when the project was last updated, in UNIX epoch time format. If the project was not updated, then createdDate and updatedDate are the same.

projectDescription_arn :: Lens' ProjectDescription (Maybe Text)

The ARN of the project.

projectDescription_placementTemplate :: Lens' ProjectDescription (Maybe PlacementTemplate)

An object describing the project's placement specifications.

projectDescription_description :: Lens' ProjectDescription (Maybe Text)

The description of the project.

projectDescription_tags :: Lens' ProjectDescription (Maybe (HashMap Text Text))

The tags (metadata key/value pairs) associated with the project.

projectDescription_projectName :: Lens' ProjectDescription Text

The name of the project for which to obtain information from.

projectDescription_createdDate :: Lens' ProjectDescription UTCTime

The date when the project was originally created, in UNIX epoch time format.

projectDescription_updatedDate :: Lens' ProjectDescription UTCTime

The date when the project was last updated, in UNIX epoch time format. If the project was not updated, then createdDate and updatedDate are the same.

ProjectSummary

data ProjectSummary

An object providing summary information for a particular project for an associated AWS account and region.

See: newProjectSummary smart constructor.

Constructors

ProjectSummary'

Fields

arn :: Maybe Text

The ARN of the project.

tags :: Maybe (HashMap Text Text)

The tags (metadata key/value pairs) associated with the project.

projectName :: Text

The name of the project being summarized.

createdDate :: POSIX

The date when the project was originally created, in UNIX epoch time format.

updatedDate :: POSIX

The date when the project was last updated, in UNIX epoch time format. If the project was not updated, then createdDate and updatedDate are the same.

newProjectSummary

Arguments

:: Text
$sel:projectName:ProjectSummary'

-> UTCTime
$sel:createdDate:ProjectSummary'

-> UTCTime
$sel:updatedDate:ProjectSummary'

-> ProjectSummary

Create a value of ProjectSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:ProjectSummary', projectSummary_arn - The ARN of the project.

$sel:tags:ProjectSummary', projectSummary_tags - The tags (metadata key/value pairs) associated with the project.

$sel:projectName:ProjectSummary', projectSummary_projectName - The name of the project being summarized.

$sel:createdDate:ProjectSummary', projectSummary_createdDate - The date when the project was originally created, in UNIX epoch time format.

$sel:updatedDate:ProjectSummary', projectSummary_updatedDate - The date when the project was last updated, in UNIX epoch time format. If the project was not updated, then createdDate and updatedDate are the same.

projectSummary_arn :: Lens' ProjectSummary (Maybe Text)

The ARN of the project.

projectSummary_tags :: Lens' ProjectSummary (Maybe (HashMap Text Text))

The tags (metadata key/value pairs) associated with the project.

projectSummary_projectName :: Lens' ProjectSummary Text

The name of the project being summarized.

projectSummary_createdDate :: Lens' ProjectSummary UTCTime

The date when the project was originally created, in UNIX epoch time format.

projectSummary_updatedDate :: Lens' ProjectSummary UTCTime

The date when the project was last updated, in UNIX epoch time format. If the project was not updated, then createdDate and updatedDate are the same.