Network.AWS.IoT1ClickProjects.Types.PlacementSummary

Description

Documentation

data PlacementSummary

An object providing summary information for a particular placement.

See: newPlacementSummary smart constructor.

Constructors

PlacementSummary'

Fields

projectName :: Text

The name of the project containing the placement.

placementName :: Text

The name of the placement being summarized.

createdDate :: POSIX

The date when the placement was originally created, in UNIX epoch time format.

updatedDate :: POSIX

The date when the placement was last updated, in UNIX epoch time format. If the placement was not updated, then createdDate and updatedDate are the same.

newPlacementSummary

Arguments

:: Text
$sel:projectName:PlacementSummary'

-> Text
$sel:placementName:PlacementSummary'

-> UTCTime
$sel:createdDate:PlacementSummary'

-> UTCTime
$sel:updatedDate:PlacementSummary'

-> PlacementSummary

Create a value of PlacementSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectName:PlacementSummary', placementSummary_projectName - The name of the project containing the placement.

$sel:placementName:PlacementSummary', placementSummary_placementName - The name of the placement being summarized.

$sel:createdDate:PlacementSummary', placementSummary_createdDate - The date when the placement was originally created, in UNIX epoch time format.

$sel:updatedDate:PlacementSummary', placementSummary_updatedDate - The date when the placement was last updated, in UNIX epoch time format. If the placement was not updated, then createdDate and updatedDate are the same.

placementSummary_projectName :: Lens' PlacementSummary Text

The name of the project containing the placement.

placementSummary_placementName :: Lens' PlacementSummary Text

The name of the placement being summarized.

placementSummary_createdDate :: Lens' PlacementSummary UTCTime

The date when the placement was originally created, in UNIX epoch time format.

placementSummary_updatedDate :: Lens' PlacementSummary UTCTime

The date when the placement was last updated, in UNIX epoch time format. If the placement was not updated, then createdDate and updatedDate are the same.