Network.AWS.IoT1ClickProjects.Types.ProjectDescription

Description

Documentation

data ProjectDescription

An object providing detailed information for a particular project associated with an AWS account and region.

See: newProjectDescription smart constructor.

Constructors

ProjectDescription'

Fields

arn :: Maybe Text

The ARN of the project.

placementTemplate :: Maybe PlacementTemplate

An object describing the project's placement specifications.

description :: Maybe Text

The description of the project.

tags :: Maybe (HashMap Text Text)

The tags (metadata key/value pairs) associated with the project.

projectName :: Text

The name of the project for which to obtain information from.

createdDate :: POSIX

The date when the project was originally created, in UNIX epoch time format.

updatedDate :: POSIX

The date when the project was last updated, in UNIX epoch time format. If the project was not updated, then createdDate and updatedDate are the same.

newProjectDescription

Arguments

:: Text
$sel:projectName:ProjectDescription'

-> UTCTime
$sel:createdDate:ProjectDescription'

-> UTCTime
$sel:updatedDate:ProjectDescription'

-> ProjectDescription

Create a value of ProjectDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:ProjectDescription', projectDescription_arn - The ARN of the project.

$sel:placementTemplate:ProjectDescription', projectDescription_placementTemplate - An object describing the project's placement specifications.

$sel:description:ProjectDescription', projectDescription_description - The description of the project.

$sel:tags:ProjectDescription', projectDescription_tags - The tags (metadata key/value pairs) associated with the project.

$sel:projectName:ProjectDescription', projectDescription_projectName - The name of the project for which to obtain information from.

$sel:createdDate:ProjectDescription', projectDescription_createdDate - The date when the project was originally created, in UNIX epoch time format.

$sel:updatedDate:ProjectDescription', projectDescription_updatedDate - The date when the project was last updated, in UNIX epoch time format. If the project was not updated, then createdDate and updatedDate are the same.

projectDescription_arn :: Lens' ProjectDescription (Maybe Text)

The ARN of the project.

projectDescription_placementTemplate :: Lens' ProjectDescription (Maybe PlacementTemplate)

An object describing the project's placement specifications.

projectDescription_description :: Lens' ProjectDescription (Maybe Text)

The description of the project.

projectDescription_tags :: Lens' ProjectDescription (Maybe (HashMap Text Text))

The tags (metadata key/value pairs) associated with the project.

projectDescription_projectName :: Lens' ProjectDescription Text

The name of the project for which to obtain information from.

projectDescription_createdDate :: Lens' ProjectDescription UTCTime

The date when the project was originally created, in UNIX epoch time format.

projectDescription_updatedDate :: Lens' ProjectDescription UTCTime

The date when the project was last updated, in UNIX epoch time format. If the project was not updated, then createdDate and updatedDate are the same.