Network.AWS.IoT1ClickProjects.Types.ProjectSummary

Description

Documentation

data ProjectSummary

An object providing summary information for a particular project for an associated AWS account and region.

See: newProjectSummary smart constructor.

Constructors

ProjectSummary'

Fields

arn :: Maybe Text

The ARN of the project.

tags :: Maybe (HashMap Text Text)

The tags (metadata key/value pairs) associated with the project.

projectName :: Text

The name of the project being summarized.

createdDate :: POSIX

The date when the project was originally created, in UNIX epoch time format.

updatedDate :: POSIX

The date when the project was last updated, in UNIX epoch time format. If the project was not updated, then createdDate and updatedDate are the same.

newProjectSummary

Arguments

:: Text
$sel:projectName:ProjectSummary'

-> UTCTime
$sel:createdDate:ProjectSummary'

-> UTCTime
$sel:updatedDate:ProjectSummary'

-> ProjectSummary

Create a value of ProjectSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:ProjectSummary', projectSummary_arn - The ARN of the project.

$sel:tags:ProjectSummary', projectSummary_tags - The tags (metadata key/value pairs) associated with the project.

$sel:projectName:ProjectSummary', projectSummary_projectName - The name of the project being summarized.

$sel:createdDate:ProjectSummary', projectSummary_createdDate - The date when the project was originally created, in UNIX epoch time format.

$sel:updatedDate:ProjectSummary', projectSummary_updatedDate - The date when the project was last updated, in UNIX epoch time format. If the project was not updated, then createdDate and updatedDate are the same.

projectSummary_arn :: Lens' ProjectSummary (Maybe Text)

The ARN of the project.

projectSummary_tags :: Lens' ProjectSummary (Maybe (HashMap Text Text))

The tags (metadata key/value pairs) associated with the project.

projectSummary_projectName :: Lens' ProjectSummary Text

The name of the project being summarized.

projectSummary_createdDate :: Lens' ProjectSummary UTCTime

The date when the project was originally created, in UNIX epoch time format.

projectSummary_updatedDate :: Lens' ProjectSummary UTCTime

The date when the project was last updated, in UNIX epoch time format. If the project was not updated, then createdDate and updatedDate are the same.