Network.AWS.IoTAnalytics.CreateChannel

Description

Used to create a channel. A channel collects data from an MQTT topic and archives the raw, unprocessed messages before publishing the data to a pipeline.

Creating a Request

data CreateChannel

See: newCreateChannel smart constructor.

Constructors

CreateChannel'

Fields

retentionPeriod :: Maybe RetentionPeriod

How long, in days, message data is kept for the channel. When customerManagedS3 storage is selected, this parameter is ignored.

tags :: Maybe (NonEmpty Tag)

Metadata which can be used to manage the channel.

channelStorage :: Maybe ChannelStorage

Where channel data is stored. You can choose one of serviceManagedS3 or customerManagedS3 storage. If not specified, the default is serviceManagedS3. You can't change this storage option after the channel is created.

channelName :: Text

The name of the channel.

newCreateChannel

Arguments

:: Text
$sel:channelName:CreateChannel'

-> CreateChannel

Create a value of CreateChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retentionPeriod:CreateChannel', createChannel_retentionPeriod - How long, in days, message data is kept for the channel. When customerManagedS3 storage is selected, this parameter is ignored.

$sel:tags:CreateChannel', createChannel_tags - Metadata which can be used to manage the channel.

$sel:channelStorage:CreateChannel', createChannel_channelStorage - Where channel data is stored. You can choose one of serviceManagedS3 or customerManagedS3 storage. If not specified, the default is serviceManagedS3. You can't change this storage option after the channel is created.

$sel:channelName:CreateChannel', createChannel_channelName - The name of the channel.

Request Lenses

createChannel_retentionPeriod :: Lens' CreateChannel (Maybe RetentionPeriod)

How long, in days, message data is kept for the channel. When customerManagedS3 storage is selected, this parameter is ignored.

createChannel_tags :: Lens' CreateChannel (Maybe (NonEmpty Tag))

Metadata which can be used to manage the channel.

createChannel_channelStorage :: Lens' CreateChannel (Maybe ChannelStorage)

Where channel data is stored. You can choose one of serviceManagedS3 or customerManagedS3 storage. If not specified, the default is serviceManagedS3. You can't change this storage option after the channel is created.

createChannel_channelName :: Lens' CreateChannel Text

The name of the channel.

Destructuring the Response

data CreateChannelResponse

See: newCreateChannelResponse smart constructor.

Constructors

CreateChannelResponse'

Fields

channelName :: Maybe Text

The name of the channel.

retentionPeriod :: Maybe RetentionPeriod

How long, in days, message data is kept for the channel.

channelArn :: Maybe Text

The ARN of the channel.

httpStatus :: Int

The response's http status code.

newCreateChannelResponse

Arguments

:: Int
$sel:httpStatus:CreateChannelResponse'

-> CreateChannelResponse

Create a value of CreateChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelName:CreateChannel', createChannelResponse_channelName - The name of the channel.

$sel:retentionPeriod:CreateChannel', createChannelResponse_retentionPeriod - How long, in days, message data is kept for the channel.

$sel:channelArn:CreateChannelResponse', createChannelResponse_channelArn - The ARN of the channel.

$sel:httpStatus:CreateChannelResponse', createChannelResponse_httpStatus - The response's http status code.

Response Lenses

createChannelResponse_channelName :: Lens' CreateChannelResponse (Maybe Text)

The name of the channel.

createChannelResponse_retentionPeriod :: Lens' CreateChannelResponse (Maybe RetentionPeriod)

How long, in days, message data is kept for the channel.

createChannelResponse_channelArn :: Lens' CreateChannelResponse (Maybe Text)

The ARN of the channel.

createChannelResponse_httpStatus :: Lens' CreateChannelResponse Int

The response's http status code.