Network.AWS.IoTAnalytics.CreateDataset

Description

Used to create a dataset. A dataset stores data retrieved from a data store by applying a queryAction (a SQL query) or a containerAction (executing a containerized application). This operation creates the skeleton of a dataset. The dataset can be populated manually by calling CreateDatasetContent or automatically according to a trigger you specify.

Creating a Request

data CreateDataset

See: newCreateDataset smart constructor.

Constructors

CreateDataset'

Fields

versioningConfiguration :: Maybe VersioningConfiguration

Optional. How many versions of dataset contents are kept. If not specified or set to null, only the latest version plus the latest succeeded version (if they are different) are kept for the time period specified by the retentionPeriod parameter. For more information, see Keeping Multiple Versions of IoT Analytics datasets in the IoT Analytics User Guide.

triggers :: Maybe [DatasetTrigger]

A list of triggers. A trigger causes dataset contents to be populated at a specified time interval or when another dataset's contents are created. The list of triggers can be empty or contain up to five DataSetTrigger objects.

retentionPeriod :: Maybe RetentionPeriod

Optional. How long, in days, versions of dataset contents are kept for the dataset. If not specified or set to null, versions of dataset contents are retained for at most 90 days. The number of versions of dataset contents retained is determined by the versioningConfiguration parameter. For more information, see Keeping Multiple Versions of IoT Analytics datasets in the IoT Analytics User Guide.

lateDataRules :: Maybe (NonEmpty LateDataRule)

A list of data rules that send notifications to CloudWatch, when data arrives late. To specify lateDataRules, the dataset must use a DeltaTimer filter.

contentDeliveryRules :: Maybe [DatasetContentDeliveryRule]

When dataset contents are created, they are delivered to destinations specified here.

tags :: Maybe (NonEmpty Tag)

Metadata which can be used to manage the dataset.

datasetName :: Text

The name of the dataset.

actions :: NonEmpty DatasetAction

A list of actions that create the dataset contents.

newCreateDataset

Arguments

:: Text
$sel:datasetName:CreateDataset'

-> NonEmpty DatasetAction
$sel:actions:CreateDataset'

-> CreateDataset

Create a value of CreateDataset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versioningConfiguration:CreateDataset', createDataset_versioningConfiguration - Optional. How many versions of dataset contents are kept. If not specified or set to null, only the latest version plus the latest succeeded version (if they are different) are kept for the time period specified by the retentionPeriod parameter. For more information, see Keeping Multiple Versions of IoT Analytics datasets in the IoT Analytics User Guide.

$sel:triggers:CreateDataset', createDataset_triggers - A list of triggers. A trigger causes dataset contents to be populated at a specified time interval or when another dataset's contents are created. The list of triggers can be empty or contain up to five DataSetTrigger objects.

$sel:retentionPeriod:CreateDataset', createDataset_retentionPeriod - Optional. How long, in days, versions of dataset contents are kept for the dataset. If not specified or set to null, versions of dataset contents are retained for at most 90 days. The number of versions of dataset contents retained is determined by the versioningConfiguration parameter. For more information, see Keeping Multiple Versions of IoT Analytics datasets in the IoT Analytics User Guide.

$sel:lateDataRules:CreateDataset', createDataset_lateDataRules - A list of data rules that send notifications to CloudWatch, when data arrives late. To specify lateDataRules, the dataset must use a DeltaTimer filter.

$sel:contentDeliveryRules:CreateDataset', createDataset_contentDeliveryRules - When dataset contents are created, they are delivered to destinations specified here.

$sel:tags:CreateDataset', createDataset_tags - Metadata which can be used to manage the dataset.

$sel:datasetName:CreateDataset', createDataset_datasetName - The name of the dataset.

$sel:actions:CreateDataset', createDataset_actions - A list of actions that create the dataset contents.

Request Lenses

createDataset_versioningConfiguration :: Lens' CreateDataset (Maybe VersioningConfiguration)

Optional. How many versions of dataset contents are kept. If not specified or set to null, only the latest version plus the latest succeeded version (if they are different) are kept for the time period specified by the retentionPeriod parameter. For more information, see Keeping Multiple Versions of IoT Analytics datasets in the IoT Analytics User Guide.

createDataset_triggers :: Lens' CreateDataset (Maybe [DatasetTrigger])

A list of triggers. A trigger causes dataset contents to be populated at a specified time interval or when another dataset's contents are created. The list of triggers can be empty or contain up to five DataSetTrigger objects.

createDataset_retentionPeriod :: Lens' CreateDataset (Maybe RetentionPeriod)

Optional. How long, in days, versions of dataset contents are kept for the dataset. If not specified or set to null, versions of dataset contents are retained for at most 90 days. The number of versions of dataset contents retained is determined by the versioningConfiguration parameter. For more information, see Keeping Multiple Versions of IoT Analytics datasets in the IoT Analytics User Guide.

createDataset_lateDataRules :: Lens' CreateDataset (Maybe (NonEmpty LateDataRule))

A list of data rules that send notifications to CloudWatch, when data arrives late. To specify lateDataRules, the dataset must use a DeltaTimer filter.

createDataset_contentDeliveryRules :: Lens' CreateDataset (Maybe [DatasetContentDeliveryRule])

When dataset contents are created, they are delivered to destinations specified here.

createDataset_tags :: Lens' CreateDataset (Maybe (NonEmpty Tag))

Metadata which can be used to manage the dataset.

createDataset_datasetName :: Lens' CreateDataset Text

The name of the dataset.

createDataset_actions :: Lens' CreateDataset (NonEmpty DatasetAction)

A list of actions that create the dataset contents.

Destructuring the Response

data CreateDatasetResponse

See: newCreateDatasetResponse smart constructor.

Constructors

CreateDatasetResponse'

Fields

datasetArn :: Maybe Text

The ARN of the dataset.

retentionPeriod :: Maybe RetentionPeriod

How long, in days, dataset contents are kept for the dataset.

datasetName :: Maybe Text

The name of the dataset.

httpStatus :: Int

The response's http status code.

newCreateDatasetResponse

Arguments

:: Int
$sel:httpStatus:CreateDatasetResponse'

-> CreateDatasetResponse

Create a value of CreateDatasetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetArn:CreateDatasetResponse', createDatasetResponse_datasetArn - The ARN of the dataset.

$sel:retentionPeriod:CreateDataset', createDatasetResponse_retentionPeriod - How long, in days, dataset contents are kept for the dataset.

$sel:datasetName:CreateDataset', createDatasetResponse_datasetName - The name of the dataset.

$sel:httpStatus:CreateDatasetResponse', createDatasetResponse_httpStatus - The response's http status code.

Response Lenses

createDatasetResponse_datasetArn :: Lens' CreateDatasetResponse (Maybe Text)

The ARN of the dataset.

createDatasetResponse_retentionPeriod :: Lens' CreateDatasetResponse (Maybe RetentionPeriod)

How long, in days, dataset contents are kept for the dataset.

createDatasetResponse_datasetName :: Lens' CreateDatasetResponse (Maybe Text)

The name of the dataset.

createDatasetResponse_httpStatus :: Lens' CreateDatasetResponse Int

The response's http status code.