Network.AWS.IoTAnalytics.CreateDatastore

Description

Creates a data store, which is a repository for messages.

Creating a Request

data CreateDatastore

See: newCreateDatastore smart constructor.

Constructors

CreateDatastore'

Fields

datastoreStorage :: Maybe DatastoreStorage

Where data in a data store is stored.. You can choose serviceManagedS3 storage, customerManagedS3 storage, or iotSiteWiseMultiLayerStorage storage. The default is serviceManagedS3. You can't change the choice of Amazon S3 storage after your data store is created.

datastorePartitions :: Maybe DatastorePartitions

Contains information about the partition dimensions in a data store.

fileFormatConfiguration :: Maybe FileFormatConfiguration

Contains the configuration information of file formats. IoT Analytics data stores support JSON and Parquet.

The default file format is JSON. You can specify only one format.

You can't change the file format after you create the data store.

retentionPeriod :: Maybe RetentionPeriod

How long, in days, message data is kept for the data store. When customerManagedS3 storage is selected, this parameter is ignored.

tags :: Maybe (NonEmpty Tag)

Metadata which can be used to manage the data store.

datastoreName :: Text

The name of the data store.

newCreateDatastore

Arguments

:: Text
$sel:datastoreName:CreateDatastore'

-> CreateDatastore

Create a value of CreateDatastore with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datastoreStorage:CreateDatastore', createDatastore_datastoreStorage - Where data in a data store is stored.. You can choose serviceManagedS3 storage, customerManagedS3 storage, or iotSiteWiseMultiLayerStorage storage. The default is serviceManagedS3. You can't change the choice of Amazon S3 storage after your data store is created.

$sel:datastorePartitions:CreateDatastore', createDatastore_datastorePartitions - Contains information about the partition dimensions in a data store.

$sel:fileFormatConfiguration:CreateDatastore', createDatastore_fileFormatConfiguration - Contains the configuration information of file formats. IoT Analytics data stores support JSON and Parquet.

The default file format is JSON. You can specify only one format.

You can't change the file format after you create the data store.

$sel:retentionPeriod:CreateDatastore', createDatastore_retentionPeriod - How long, in days, message data is kept for the data store. When customerManagedS3 storage is selected, this parameter is ignored.

$sel:tags:CreateDatastore', createDatastore_tags - Metadata which can be used to manage the data store.

$sel:datastoreName:CreateDatastore', createDatastore_datastoreName - The name of the data store.

Request Lenses

createDatastore_datastoreStorage :: Lens' CreateDatastore (Maybe DatastoreStorage)

Where data in a data store is stored.. You can choose serviceManagedS3 storage, customerManagedS3 storage, or iotSiteWiseMultiLayerStorage storage. The default is serviceManagedS3. You can't change the choice of Amazon S3 storage after your data store is created.

createDatastore_datastorePartitions :: Lens' CreateDatastore (Maybe DatastorePartitions)

Contains information about the partition dimensions in a data store.

createDatastore_fileFormatConfiguration :: Lens' CreateDatastore (Maybe FileFormatConfiguration)

Contains the configuration information of file formats. IoT Analytics data stores support JSON and Parquet.

The default file format is JSON. You can specify only one format.

You can't change the file format after you create the data store.

createDatastore_retentionPeriod :: Lens' CreateDatastore (Maybe RetentionPeriod)

How long, in days, message data is kept for the data store. When customerManagedS3 storage is selected, this parameter is ignored.

createDatastore_tags :: Lens' CreateDatastore (Maybe (NonEmpty Tag))

Metadata which can be used to manage the data store.

createDatastore_datastoreName :: Lens' CreateDatastore Text

The name of the data store.

Destructuring the Response

data CreateDatastoreResponse

See: newCreateDatastoreResponse smart constructor.

Constructors

CreateDatastoreResponse'

Fields

datastoreArn :: Maybe Text

The ARN of the data store.

retentionPeriod :: Maybe RetentionPeriod

How long, in days, message data is kept for the data store.

datastoreName :: Maybe Text

The name of the data store.

httpStatus :: Int

The response's http status code.

newCreateDatastoreResponse

Arguments

:: Int
$sel:httpStatus:CreateDatastoreResponse'

-> CreateDatastoreResponse

Create a value of CreateDatastoreResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datastoreArn:CreateDatastoreResponse', createDatastoreResponse_datastoreArn - The ARN of the data store.

$sel:retentionPeriod:CreateDatastore', createDatastoreResponse_retentionPeriod - How long, in days, message data is kept for the data store.

$sel:datastoreName:CreateDatastore', createDatastoreResponse_datastoreName - The name of the data store.

$sel:httpStatus:CreateDatastoreResponse', createDatastoreResponse_httpStatus - The response's http status code.

Response Lenses

createDatastoreResponse_datastoreArn :: Lens' CreateDatastoreResponse (Maybe Text)

The ARN of the data store.

createDatastoreResponse_retentionPeriod :: Lens' CreateDatastoreResponse (Maybe RetentionPeriod)

How long, in days, message data is kept for the data store.

createDatastoreResponse_datastoreName :: Lens' CreateDatastoreResponse (Maybe Text)

The name of the data store.

createDatastoreResponse_httpStatus :: Lens' CreateDatastoreResponse Int

The response's http status code.