Network.AWS.IoTAnalytics.DeleteDataset

Description

Deletes the specified dataset.

You do not have to delete the content of the dataset before you perform this operation.

Creating a Request

data DeleteDataset

See: newDeleteDataset smart constructor.

Constructors

DeleteDataset'

Fields

datasetName :: Text

The name of the dataset to delete.

newDeleteDataset

Arguments

:: Text
$sel:datasetName:DeleteDataset'

-> DeleteDataset

Create a value of DeleteDataset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetName:DeleteDataset', deleteDataset_datasetName - The name of the dataset to delete.

Request Lenses

deleteDataset_datasetName :: Lens' DeleteDataset Text

The name of the dataset to delete.

Destructuring the Response

data DeleteDatasetResponse

See: newDeleteDatasetResponse smart constructor.

Constructors

DeleteDatasetResponse'

newDeleteDatasetResponse :: DeleteDatasetResponse

Create a value of DeleteDatasetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.