Network.AWS.IoTAnalytics.GetDatasetContent

Description

Retrieves the contents of a dataset as presigned URIs.

Creating a Request

data GetDatasetContent

See: newGetDatasetContent smart constructor.

Constructors

GetDatasetContent'

Fields

versionId :: Maybe Text

The version of the dataset whose contents are retrieved. You can also use the strings "$LATEST" or "$LATEST_SUCCEEDED" to retrieve the contents of the latest or latest successfully completed dataset. If not specified, "$LATEST_SUCCEEDED" is the default.

datasetName :: Text

The name of the dataset whose contents are retrieved.

newGetDatasetContent

Arguments

:: Text
$sel:datasetName:GetDatasetContent'

-> GetDatasetContent

Create a value of GetDatasetContent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionId:GetDatasetContent', getDatasetContent_versionId - The version of the dataset whose contents are retrieved. You can also use the strings "$LATEST" or "$LATEST_SUCCEEDED" to retrieve the contents of the latest or latest successfully completed dataset. If not specified, "$LATEST_SUCCEEDED" is the default.

$sel:datasetName:GetDatasetContent', getDatasetContent_datasetName - The name of the dataset whose contents are retrieved.

Request Lenses

getDatasetContent_versionId :: Lens' GetDatasetContent (Maybe Text)

The version of the dataset whose contents are retrieved. You can also use the strings "$LATEST" or "$LATEST_SUCCEEDED" to retrieve the contents of the latest or latest successfully completed dataset. If not specified, "$LATEST_SUCCEEDED" is the default.

getDatasetContent_datasetName :: Lens' GetDatasetContent Text

The name of the dataset whose contents are retrieved.

Destructuring the Response

data GetDatasetContentResponse

See: newGetDatasetContentResponse smart constructor.

Constructors

GetDatasetContentResponse'

Fields

status :: Maybe DatasetContentStatus

The status of the dataset content.

entries :: Maybe [DatasetEntry]

A list of DatasetEntry objects.

timestamp :: Maybe POSIX

The time when the request was made.

httpStatus :: Int

The response's http status code.

newGetDatasetContentResponse

Arguments

:: Int
$sel:httpStatus:GetDatasetContentResponse'

-> GetDatasetContentResponse

Create a value of GetDatasetContentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetDatasetContentResponse', getDatasetContentResponse_status - The status of the dataset content.

$sel:entries:GetDatasetContentResponse', getDatasetContentResponse_entries - A list of DatasetEntry objects.

$sel:timestamp:GetDatasetContentResponse', getDatasetContentResponse_timestamp - The time when the request was made.

$sel:httpStatus:GetDatasetContentResponse', getDatasetContentResponse_httpStatus - The response's http status code.

Response Lenses

getDatasetContentResponse_status :: Lens' GetDatasetContentResponse (Maybe DatasetContentStatus)

The status of the dataset content.

getDatasetContentResponse_entries :: Lens' GetDatasetContentResponse (Maybe [DatasetEntry])

A list of DatasetEntry objects.

getDatasetContentResponse_timestamp :: Lens' GetDatasetContentResponse (Maybe UTCTime)

The time when the request was made.

getDatasetContentResponse_httpStatus :: Lens' GetDatasetContentResponse Int

The response's http status code.