Network.AWS.IoTAnalytics.ListDatasets

Description

Retrieves information about datasets.

This operation returns paginated results.

Creating a Request

data ListDatasets

See: newListDatasets smart constructor.

Constructors

ListDatasets'

Fields

nextToken :: Maybe Text

The token for the next set of results.

maxResults :: Maybe Natural

The maximum number of results to return in this request.

The default value is 100.

newListDatasets :: ListDatasets

Create a value of ListDatasets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDatasets', listDatasets_nextToken - The token for the next set of results.

$sel:maxResults:ListDatasets', listDatasets_maxResults - The maximum number of results to return in this request.

The default value is 100.

Request Lenses

listDatasets_nextToken :: Lens' ListDatasets (Maybe Text)

The token for the next set of results.

listDatasets_maxResults :: Lens' ListDatasets (Maybe Natural)

The maximum number of results to return in this request.

The default value is 100.

Destructuring the Response

data ListDatasetsResponse

See: newListDatasetsResponse smart constructor.

Constructors

ListDatasetsResponse'

Fields

nextToken :: Maybe Text

The token to retrieve the next set of results, or null if there are no more results.

datasetSummaries :: Maybe [DatasetSummary]

A list of DatasetSummary objects.

httpStatus :: Int

The response's http status code.

newListDatasetsResponse

Arguments

:: Int
$sel:httpStatus:ListDatasetsResponse'

-> ListDatasetsResponse

Create a value of ListDatasetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDatasets', listDatasetsResponse_nextToken - The token to retrieve the next set of results, or null if there are no more results.

$sel:datasetSummaries:ListDatasetsResponse', listDatasetsResponse_datasetSummaries - A list of DatasetSummary objects.

$sel:httpStatus:ListDatasetsResponse', listDatasetsResponse_httpStatus - The response's http status code.

Response Lenses

listDatasetsResponse_nextToken :: Lens' ListDatasetsResponse (Maybe Text)

The token to retrieve the next set of results, or null if there are no more results.

listDatasetsResponse_datasetSummaries :: Lens' ListDatasetsResponse (Maybe [DatasetSummary])

A list of DatasetSummary objects.

listDatasetsResponse_httpStatus :: Lens' ListDatasetsResponse Int

The response's http status code.