Network.AWS.IoTAnalytics.ListDatastores

Description

Retrieves a list of data stores.

This operation returns paginated results.

Creating a Request

data ListDatastores

See: newListDatastores smart constructor.

Constructors

ListDatastores'

Fields

nextToken :: Maybe Text

The token for the next set of results.

maxResults :: Maybe Natural

The maximum number of results to return in this request.

The default value is 100.

newListDatastores :: ListDatastores

Create a value of ListDatastores with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDatastores', listDatastores_nextToken - The token for the next set of results.

$sel:maxResults:ListDatastores', listDatastores_maxResults - The maximum number of results to return in this request.

The default value is 100.

Request Lenses

listDatastores_nextToken :: Lens' ListDatastores (Maybe Text)

The token for the next set of results.

listDatastores_maxResults :: Lens' ListDatastores (Maybe Natural)

The maximum number of results to return in this request.

The default value is 100.

Destructuring the Response

data ListDatastoresResponse

See: newListDatastoresResponse smart constructor.

Constructors

ListDatastoresResponse'

Fields

nextToken :: Maybe Text

The token to retrieve the next set of results, or null if there are no more results.

datastoreSummaries :: Maybe [DatastoreSummary]

A list of DatastoreSummary objects.

httpStatus :: Int

The response's http status code.

newListDatastoresResponse

Arguments

:: Int
$sel:httpStatus:ListDatastoresResponse'

-> ListDatastoresResponse

Create a value of ListDatastoresResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDatastores', listDatastoresResponse_nextToken - The token to retrieve the next set of results, or null if there are no more results.

$sel:datastoreSummaries:ListDatastoresResponse', listDatastoresResponse_datastoreSummaries - A list of DatastoreSummary objects.

$sel:httpStatus:ListDatastoresResponse', listDatastoresResponse_httpStatus - The response's http status code.

Response Lenses

listDatastoresResponse_nextToken :: Lens' ListDatastoresResponse (Maybe Text)

The token to retrieve the next set of results, or null if there are no more results.

listDatastoresResponse_datastoreSummaries :: Lens' ListDatastoresResponse (Maybe [DatastoreSummary])

A list of DatastoreSummary objects.

listDatastoresResponse_httpStatus :: Lens' ListDatastoresResponse Int

The response's http status code.