Network.AWS.IoTAnalytics.StartPipelineReprocessing

Description

Starts the reprocessing of raw message data through the pipeline.

Creating a Request

data StartPipelineReprocessing

See: newStartPipelineReprocessing smart constructor.

Constructors

StartPipelineReprocessing'

Fields

channelMessages :: Maybe ChannelMessages

Specifies one or more sets of channel messages that you want to reprocess.

If you use the channelMessages object, you must not specify a value for startTime and endTime.

startTime :: Maybe POSIX

The start time (inclusive) of raw message data that is reprocessed.

If you specify a value for the startTime parameter, you must not use the channelMessages object.

endTime :: Maybe POSIX

The end time (exclusive) of raw message data that is reprocessed.

If you specify a value for the endTime parameter, you must not use the channelMessages object.

pipelineName :: Text

The name of the pipeline on which to start reprocessing.

newStartPipelineReprocessing

Arguments

:: Text
$sel:pipelineName:StartPipelineReprocessing'

-> StartPipelineReprocessing

Create a value of StartPipelineReprocessing with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelMessages:StartPipelineReprocessing', startPipelineReprocessing_channelMessages - Specifies one or more sets of channel messages that you want to reprocess.

If you use the channelMessages object, you must not specify a value for startTime and endTime.

$sel:startTime:StartPipelineReprocessing', startPipelineReprocessing_startTime - The start time (inclusive) of raw message data that is reprocessed.

If you specify a value for the startTime parameter, you must not use the channelMessages object.

$sel:endTime:StartPipelineReprocessing', startPipelineReprocessing_endTime - The end time (exclusive) of raw message data that is reprocessed.

If you specify a value for the endTime parameter, you must not use the channelMessages object.

$sel:pipelineName:StartPipelineReprocessing', startPipelineReprocessing_pipelineName - The name of the pipeline on which to start reprocessing.

Request Lenses

startPipelineReprocessing_channelMessages :: Lens' StartPipelineReprocessing (Maybe ChannelMessages)

Specifies one or more sets of channel messages that you want to reprocess.

If you use the channelMessages object, you must not specify a value for startTime and endTime.

startPipelineReprocessing_startTime :: Lens' StartPipelineReprocessing (Maybe UTCTime)

The start time (inclusive) of raw message data that is reprocessed.

If you specify a value for the startTime parameter, you must not use the channelMessages object.

startPipelineReprocessing_endTime :: Lens' StartPipelineReprocessing (Maybe UTCTime)

The end time (exclusive) of raw message data that is reprocessed.

If you specify a value for the endTime parameter, you must not use the channelMessages object.

startPipelineReprocessing_pipelineName :: Lens' StartPipelineReprocessing Text

The name of the pipeline on which to start reprocessing.

Destructuring the Response

data StartPipelineReprocessingResponse

See: newStartPipelineReprocessingResponse smart constructor.

Constructors

StartPipelineReprocessingResponse'

Fields

reprocessingId :: Maybe Text

The ID of the pipeline reprocessing activity that was started.

httpStatus :: Int

The response's http status code.

newStartPipelineReprocessingResponse

Arguments

:: Int
$sel:httpStatus:StartPipelineReprocessingResponse'

-> StartPipelineReprocessingResponse

Create a value of StartPipelineReprocessingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reprocessingId:StartPipelineReprocessingResponse', startPipelineReprocessingResponse_reprocessingId - The ID of the pipeline reprocessing activity that was started.

$sel:httpStatus:StartPipelineReprocessingResponse', startPipelineReprocessingResponse_httpStatus - The response's http status code.

Response Lenses

startPipelineReprocessingResponse_reprocessingId :: Lens' StartPipelineReprocessingResponse (Maybe Text)

The ID of the pipeline reprocessing activity that was started.

startPipelineReprocessingResponse_httpStatus :: Lens' StartPipelineReprocessingResponse Int

The response's http status code.