Network.AWS.IoTAnalytics.Types.Channel

Description

Documentation

data Channel

A collection of data from an MQTT topic. Channels archive the raw, unprocessed messages before publishing the data to a pipeline.

See: newChannel smart constructor.

Constructors

Channel'

Fields

lastMessageArrivalTime :: Maybe POSIX

The last time when a new message arrived in the channel.

IoT Analytics updates this value at most once per minute for one channel. Hence, the lastMessageArrivalTime value is an approximation.

This feature only applies to messages that arrived in the data store after October 23, 2020.

status :: Maybe ChannelStatus

The status of the channel.

creationTime :: Maybe POSIX

When the channel was created.

lastUpdateTime :: Maybe POSIX

When the channel was last updated.

arn :: Maybe Text

The ARN of the channel.

name :: Maybe Text

The name of the channel.

retentionPeriod :: Maybe RetentionPeriod

How long, in days, message data is kept for the channel.

storage :: Maybe ChannelStorage

Where channel data is stored. You can choose one of serviceManagedS3 or customerManagedS3 storage. If not specified, the default is serviceManagedS3. You can't change this storage option after the channel is created.

newChannel :: Channel

Create a value of Channel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastMessageArrivalTime:Channel', channel_lastMessageArrivalTime - The last time when a new message arrived in the channel.

IoT Analytics updates this value at most once per minute for one channel. Hence, the lastMessageArrivalTime value is an approximation.

This feature only applies to messages that arrived in the data store after October 23, 2020.

$sel:status:Channel', channel_status - The status of the channel.

$sel:creationTime:Channel', channel_creationTime - When the channel was created.

$sel:lastUpdateTime:Channel', channel_lastUpdateTime - When the channel was last updated.

$sel:arn:Channel', channel_arn - The ARN of the channel.

$sel:name:Channel', channel_name - The name of the channel.

$sel:retentionPeriod:Channel', channel_retentionPeriod - How long, in days, message data is kept for the channel.

$sel:storage:Channel', channel_storage - Where channel data is stored. You can choose one of serviceManagedS3 or customerManagedS3 storage. If not specified, the default is serviceManagedS3. You can't change this storage option after the channel is created.

channel_lastMessageArrivalTime :: Lens' Channel (Maybe UTCTime)

The last time when a new message arrived in the channel.

IoT Analytics updates this value at most once per minute for one channel. Hence, the lastMessageArrivalTime value is an approximation.

This feature only applies to messages that arrived in the data store after October 23, 2020.

channel_status :: Lens' Channel (Maybe ChannelStatus)

The status of the channel.

channel_creationTime :: Lens' Channel (Maybe UTCTime)

When the channel was created.

channel_lastUpdateTime :: Lens' Channel (Maybe UTCTime)

When the channel was last updated.

channel_arn :: Lens' Channel (Maybe Text)

The ARN of the channel.

channel_name :: Lens' Channel (Maybe Text)

The name of the channel.

channel_retentionPeriod :: Lens' Channel (Maybe RetentionPeriod)

How long, in days, message data is kept for the channel.

channel_storage :: Lens' Channel (Maybe ChannelStorage)

Where channel data is stored. You can choose one of serviceManagedS3 or customerManagedS3 storage. If not specified, the default is serviceManagedS3. You can't change this storage option after the channel is created.