Network.AWS.IoTAnalytics.Types.ChannelSummary

Description

Documentation

data ChannelSummary

A summary of information about a channel.

See: newChannelSummary smart constructor.

Constructors

ChannelSummary'

Fields

lastMessageArrivalTime :: Maybe POSIX

The last time when a new message arrived in the channel.

IoT Analytics updates this value at most once per minute for one channel. Hence, the lastMessageArrivalTime value is an approximation.

This feature only applies to messages that arrived in the data store after October 23, 2020.

status :: Maybe ChannelStatus

The status of the channel.

channelName :: Maybe Text

The name of the channel.

creationTime :: Maybe POSIX

When the channel was created.

lastUpdateTime :: Maybe POSIX

The last time the channel was updated.

channelStorage :: Maybe ChannelStorageSummary

Where channel data is stored.

newChannelSummary :: ChannelSummary

Create a value of ChannelSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastMessageArrivalTime:ChannelSummary', channelSummary_lastMessageArrivalTime - The last time when a new message arrived in the channel.

IoT Analytics updates this value at most once per minute for one channel. Hence, the lastMessageArrivalTime value is an approximation.

This feature only applies to messages that arrived in the data store after October 23, 2020.

$sel:status:ChannelSummary', channelSummary_status - The status of the channel.

$sel:channelName:ChannelSummary', channelSummary_channelName - The name of the channel.

$sel:creationTime:ChannelSummary', channelSummary_creationTime - When the channel was created.

$sel:lastUpdateTime:ChannelSummary', channelSummary_lastUpdateTime - The last time the channel was updated.

$sel:channelStorage:ChannelSummary', channelSummary_channelStorage - Where channel data is stored.

channelSummary_lastMessageArrivalTime :: Lens' ChannelSummary (Maybe UTCTime)

The last time when a new message arrived in the channel.

IoT Analytics updates this value at most once per minute for one channel. Hence, the lastMessageArrivalTime value is an approximation.

This feature only applies to messages that arrived in the data store after October 23, 2020.

channelSummary_status :: Lens' ChannelSummary (Maybe ChannelStatus)

The status of the channel.

channelSummary_channelName :: Lens' ChannelSummary (Maybe Text)

The name of the channel.

channelSummary_creationTime :: Lens' ChannelSummary (Maybe UTCTime)

When the channel was created.

channelSummary_lastUpdateTime :: Lens' ChannelSummary (Maybe UTCTime)

The last time the channel was updated.

channelSummary_channelStorage :: Lens' ChannelSummary (Maybe ChannelStorageSummary)

Where channel data is stored.