Network.AWS.IoTAnalytics.Types.ContainerDatasetAction

Description

Documentation

data ContainerDatasetAction

Information required to run the containerAction to produce dataset contents.

See: newContainerDatasetAction smart constructor.

Constructors

ContainerDatasetAction'

Fields

variables :: Maybe [Variable]

The values of variables used in the context of the execution of the containerized application (basically, parameters passed to the application). Each variable must have a name and a value given by one of stringValue, datasetContentVersionValue, or outputFileUriValue.

image :: Text

The ARN of the Docker container stored in your account. The Docker container contains an application and required support libraries and is used to generate dataset contents.

executionRoleArn :: Text

The ARN of the role that gives permission to the system to access required resources to run the containerAction. This includes, at minimum, permission to retrieve the dataset contents that are the input to the containerized application.

resourceConfiguration :: ResourceConfiguration

Configuration of the resource that executes the containerAction.

newContainerDatasetAction

Arguments

:: Text
$sel:image:ContainerDatasetAction'

-> Text
$sel:executionRoleArn:ContainerDatasetAction'

-> ResourceConfiguration
$sel:resourceConfiguration:ContainerDatasetAction'

-> ContainerDatasetAction

Create a value of ContainerDatasetAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:variables:ContainerDatasetAction', containerDatasetAction_variables - The values of variables used in the context of the execution of the containerized application (basically, parameters passed to the application). Each variable must have a name and a value given by one of stringValue, datasetContentVersionValue, or outputFileUriValue.

$sel:image:ContainerDatasetAction', containerDatasetAction_image - The ARN of the Docker container stored in your account. The Docker container contains an application and required support libraries and is used to generate dataset contents.

$sel:executionRoleArn:ContainerDatasetAction', containerDatasetAction_executionRoleArn - The ARN of the role that gives permission to the system to access required resources to run the containerAction. This includes, at minimum, permission to retrieve the dataset contents that are the input to the containerized application.

$sel:resourceConfiguration:ContainerDatasetAction', containerDatasetAction_resourceConfiguration - Configuration of the resource that executes the containerAction.

containerDatasetAction_variables :: Lens' ContainerDatasetAction (Maybe [Variable])

The values of variables used in the context of the execution of the containerized application (basically, parameters passed to the application). Each variable must have a name and a value given by one of stringValue, datasetContentVersionValue, or outputFileUriValue.

containerDatasetAction_image :: Lens' ContainerDatasetAction Text

The ARN of the Docker container stored in your account. The Docker container contains an application and required support libraries and is used to generate dataset contents.

containerDatasetAction_executionRoleArn :: Lens' ContainerDatasetAction Text

The ARN of the role that gives permission to the system to access required resources to run the containerAction. This includes, at minimum, permission to retrieve the dataset contents that are the input to the containerized application.

containerDatasetAction_resourceConfiguration :: Lens' ContainerDatasetAction ResourceConfiguration

Configuration of the resource that executes the containerAction.