Network.AWS.IoTAnalytics.Types.Datastore

Description

Documentation

data Datastore

Information about a data store.

See: newDatastore smart constructor.

Constructors

Datastore'

Fields

lastMessageArrivalTime :: Maybe POSIX

The last time when a new message arrived in the data store.

IoT Analytics updates this value at most once per minute for Amazon Simple Storage Service one data store. Hence, the lastMessageArrivalTime value is an approximation.

This feature only applies to messages that arrived in the data store after October 23, 2020.

status :: Maybe DatastoreStatus

The status of a data store:

CREATING

The data store is being created.

ACTIVE

The data store has been created and can be used.

DELETING

The data store is being deleted.

creationTime :: Maybe POSIX

When the data store was created.

datastorePartitions :: Maybe DatastorePartitions

Contains information about the partition dimensions in a data store.

lastUpdateTime :: Maybe POSIX

The last time the data store was updated.

fileFormatConfiguration :: Maybe FileFormatConfiguration

Contains the configuration information of file formats. IoT Analytics data stores support JSON and Parquet.

The default file format is JSON. You can specify only one format.

You can't change the file format after you create the data store.

arn :: Maybe Text

The ARN of the data store.

name :: Maybe Text

The name of the data store.

retentionPeriod :: Maybe RetentionPeriod

How long, in days, message data is kept for the data store. When customerManagedS3 storage is selected, this parameter is ignored.

storage :: Maybe DatastoreStorage

Where data in a data store is stored.. You can choose serviceManagedS3 storage, customerManagedS3 storage, or iotSiteWiseMultiLayerStorage storage. The default is serviceManagedS3. You can't change the choice of Amazon S3 storage after your data store is created.

newDatastore :: Datastore

Create a value of Datastore with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastMessageArrivalTime:Datastore', datastore_lastMessageArrivalTime - The last time when a new message arrived in the data store.

IoT Analytics updates this value at most once per minute for Amazon Simple Storage Service one data store. Hence, the lastMessageArrivalTime value is an approximation.

This feature only applies to messages that arrived in the data store after October 23, 2020.

$sel:status:Datastore', datastore_status - The status of a data store:

CREATING

The data store is being created.

ACTIVE

The data store has been created and can be used.

DELETING

The data store is being deleted.

$sel:creationTime:Datastore', datastore_creationTime - When the data store was created.

$sel:datastorePartitions:Datastore', datastore_datastorePartitions - Contains information about the partition dimensions in a data store.

$sel:lastUpdateTime:Datastore', datastore_lastUpdateTime - The last time the data store was updated.

$sel:fileFormatConfiguration:Datastore', datastore_fileFormatConfiguration - Contains the configuration information of file formats. IoT Analytics data stores support JSON and Parquet.

The default file format is JSON. You can specify only one format.

You can't change the file format after you create the data store.

$sel:arn:Datastore', datastore_arn - The ARN of the data store.

$sel:name:Datastore', datastore_name - The name of the data store.

$sel:retentionPeriod:Datastore', datastore_retentionPeriod - How long, in days, message data is kept for the data store. When customerManagedS3 storage is selected, this parameter is ignored.

$sel:storage:Datastore', datastore_storage - Where data in a data store is stored.. You can choose serviceManagedS3 storage, customerManagedS3 storage, or iotSiteWiseMultiLayerStorage storage. The default is serviceManagedS3. You can't change the choice of Amazon S3 storage after your data store is created.

datastore_lastMessageArrivalTime :: Lens' Datastore (Maybe UTCTime)

The last time when a new message arrived in the data store.

IoT Analytics updates this value at most once per minute for Amazon Simple Storage Service one data store. Hence, the lastMessageArrivalTime value is an approximation.

This feature only applies to messages that arrived in the data store after October 23, 2020.

datastore_status :: Lens' Datastore (Maybe DatastoreStatus)

The status of a data store:

CREATING

The data store is being created.

ACTIVE

The data store has been created and can be used.

DELETING

The data store is being deleted.

datastore_creationTime :: Lens' Datastore (Maybe UTCTime)

When the data store was created.

datastore_datastorePartitions :: Lens' Datastore (Maybe DatastorePartitions)

Contains information about the partition dimensions in a data store.

datastore_lastUpdateTime :: Lens' Datastore (Maybe UTCTime)

The last time the data store was updated.

datastore_fileFormatConfiguration :: Lens' Datastore (Maybe FileFormatConfiguration)

Contains the configuration information of file formats. IoT Analytics data stores support JSON and Parquet.

The default file format is JSON. You can specify only one format.

You can't change the file format after you create the data store.

datastore_arn :: Lens' Datastore (Maybe Text)

The ARN of the data store.

datastore_name :: Lens' Datastore (Maybe Text)

The name of the data store.

datastore_retentionPeriod :: Lens' Datastore (Maybe RetentionPeriod)

How long, in days, message data is kept for the data store. When customerManagedS3 storage is selected, this parameter is ignored.

datastore_storage :: Lens' Datastore (Maybe DatastoreStorage)

Where data in a data store is stored.. You can choose serviceManagedS3 storage, customerManagedS3 storage, or iotSiteWiseMultiLayerStorage storage. The default is serviceManagedS3. You can't change the choice of Amazon S3 storage after your data store is created.