Network.AWS.IoTAnalytics.Types.DeltaTimeSessionWindowConfiguration

Description

Documentation

data DeltaTimeSessionWindowConfiguration

A structure that contains the configuration information of a delta time session window.

DeltaTime specifies a time interval. You can use DeltaTime to create dataset contents with data that has arrived in the data store since the last execution. For an example of DeltaTime, see Creating a SQL dataset with a delta window (CLI) in the IoT Analytics User Guide.

See: newDeltaTimeSessionWindowConfiguration smart constructor.

Constructors

DeltaTimeSessionWindowConfiguration'

Fields

timeoutInMinutes :: Natural

A time interval. You can use timeoutInMinutes so that IoT Analytics can batch up late data notifications that have been generated since the last execution. IoT Analytics sends one batch of notifications to Amazon CloudWatch Events at one time.

For more information about how to write a timestamp expression, see Date and Time Functions and Operators, in the Presto 0.172 Documentation.

newDeltaTimeSessionWindowConfiguration

Arguments

:: Natural
$sel:timeoutInMinutes:DeltaTimeSessionWindowConfiguration'

-> DeltaTimeSessionWindowConfiguration

Create a value of DeltaTimeSessionWindowConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeoutInMinutes:DeltaTimeSessionWindowConfiguration', deltaTimeSessionWindowConfiguration_timeoutInMinutes - A time interval. You can use timeoutInMinutes so that IoT Analytics can batch up late data notifications that have been generated since the last execution. IoT Analytics sends one batch of notifications to Amazon CloudWatch Events at one time.

For more information about how to write a timestamp expression, see Date and Time Functions and Operators, in the Presto 0.172 Documentation.

deltaTimeSessionWindowConfiguration_timeoutInMinutes :: Lens' DeltaTimeSessionWindowConfiguration Natural

A time interval. You can use timeoutInMinutes so that IoT Analytics can batch up late data notifications that have been generated since the last execution. IoT Analytics sends one batch of notifications to Amazon CloudWatch Events at one time.

For more information about how to write a timestamp expression, see Date and Time Functions and Operators, in the Presto 0.172 Documentation.