Network.AWS.IoTAnalytics.Types.DeviceShadowEnrichActivity

Description

Documentation

data DeviceShadowEnrichActivity

An activity that adds information from the IoT Device Shadow service to a message.

See: newDeviceShadowEnrichActivity smart constructor.

Constructors

DeviceShadowEnrichActivity'

Fields

next :: Maybe Text

The next activity in the pipeline.

name :: Text

The name of the deviceShadowEnrich activity.

attribute :: Text

The name of the attribute that is added to the message.

thingName :: Text

The name of the IoT device whose shadow information is added to the message.

roleArn :: Text

The ARN of the role that allows access to the device's shadow.

newDeviceShadowEnrichActivity

Arguments

:: Text
$sel:name:DeviceShadowEnrichActivity'

-> Text
$sel:attribute:DeviceShadowEnrichActivity'

-> Text
$sel:thingName:DeviceShadowEnrichActivity'

-> Text
$sel:roleArn:DeviceShadowEnrichActivity'

-> DeviceShadowEnrichActivity

Create a value of DeviceShadowEnrichActivity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:next:DeviceShadowEnrichActivity', deviceShadowEnrichActivity_next - The next activity in the pipeline.

$sel:name:DeviceShadowEnrichActivity', deviceShadowEnrichActivity_name - The name of the deviceShadowEnrich activity.

$sel:attribute:DeviceShadowEnrichActivity', deviceShadowEnrichActivity_attribute - The name of the attribute that is added to the message.

$sel:thingName:DeviceShadowEnrichActivity', deviceShadowEnrichActivity_thingName - The name of the IoT device whose shadow information is added to the message.

$sel:roleArn:DeviceShadowEnrichActivity', deviceShadowEnrichActivity_roleArn - The ARN of the role that allows access to the device's shadow.

deviceShadowEnrichActivity_next :: Lens' DeviceShadowEnrichActivity (Maybe Text)

The next activity in the pipeline.

deviceShadowEnrichActivity_name :: Lens' DeviceShadowEnrichActivity Text

The name of the deviceShadowEnrich activity.

deviceShadowEnrichActivity_attribute :: Lens' DeviceShadowEnrichActivity Text

The name of the attribute that is added to the message.

deviceShadowEnrichActivity_thingName :: Lens' DeviceShadowEnrichActivity Text

The name of the IoT device whose shadow information is added to the message.

deviceShadowEnrichActivity_roleArn :: Lens' DeviceShadowEnrichActivity Text

The ARN of the role that allows access to the device's shadow.