Network.AWS.IoTAnalytics.Types.GlueConfiguration

Description

Documentation

data GlueConfiguration

Configuration information for coordination with Glue, a fully managed extract, transform and load (ETL) service.

See: newGlueConfiguration smart constructor.

Constructors

GlueConfiguration'

Fields

tableName :: Text

The name of the table in your Glue Data Catalog that is used to perform the ETL operations. An Glue Data Catalog table contains partitioned data and descriptions of data sources and targets.

databaseName :: Text

The name of the database in your Glue Data Catalog in which the table is located. An Glue Data Catalog database contains metadata tables.

newGlueConfiguration

Arguments

:: Text
$sel:tableName:GlueConfiguration'

-> Text
$sel:databaseName:GlueConfiguration'

-> GlueConfiguration

Create a value of GlueConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tableName:GlueConfiguration', glueConfiguration_tableName - The name of the table in your Glue Data Catalog that is used to perform the ETL operations. An Glue Data Catalog table contains partitioned data and descriptions of data sources and targets.

$sel:databaseName:GlueConfiguration', glueConfiguration_databaseName - The name of the database in your Glue Data Catalog in which the table is located. An Glue Data Catalog database contains metadata tables.

glueConfiguration_tableName :: Lens' GlueConfiguration Text

The name of the table in your Glue Data Catalog that is used to perform the ETL operations. An Glue Data Catalog table contains partitioned data and descriptions of data sources and targets.

glueConfiguration_databaseName :: Lens' GlueConfiguration Text

The name of the database in your Glue Data Catalog in which the table is located. An Glue Data Catalog database contains metadata tables.