Network.AWS.IoTAnalytics.Types.LateDataRule

Description

Documentation

data LateDataRule

A structure that contains the name and configuration information of a late data rule.

See: newLateDataRule smart constructor.

Constructors

LateDataRule'

Fields

ruleName :: Maybe Text

The name of the late data rule.

ruleConfiguration :: LateDataRuleConfiguration

The information needed to configure the late data rule.

newLateDataRule

Arguments

:: LateDataRuleConfiguration
$sel:ruleConfiguration:LateDataRule'

-> LateDataRule

Create a value of LateDataRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleName:LateDataRule', lateDataRule_ruleName - The name of the late data rule.

$sel:ruleConfiguration:LateDataRule', lateDataRule_ruleConfiguration - The information needed to configure the late data rule.

lateDataRule_ruleName :: Lens' LateDataRule (Maybe Text)

The name of the late data rule.

lateDataRule_ruleConfiguration :: Lens' LateDataRule LateDataRuleConfiguration

The information needed to configure the late data rule.