Network.AWS.IoTAnalytics.Types.LateDataRuleConfiguration

Description

Documentation

data LateDataRuleConfiguration

The information needed to configure a delta time session window.

See: newLateDataRuleConfiguration smart constructor.

Constructors

LateDataRuleConfiguration'

Fields

deltaTimeSessionWindowConfiguration :: Maybe DeltaTimeSessionWindowConfiguration

The information needed to configure a delta time session window.

newLateDataRuleConfiguration :: LateDataRuleConfiguration

Create a value of LateDataRuleConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deltaTimeSessionWindowConfiguration:LateDataRuleConfiguration', lateDataRuleConfiguration_deltaTimeSessionWindowConfiguration - The information needed to configure a delta time session window.

lateDataRuleConfiguration_deltaTimeSessionWindowConfiguration :: Lens' LateDataRuleConfiguration (Maybe DeltaTimeSessionWindowConfiguration)

The information needed to configure a delta time session window.