Network.AWS.IoTAnalytics.Types.Pipeline

Description

Documentation

data Pipeline

Contains information about a pipeline.

See: newPipeline smart constructor.

Constructors

Pipeline'

Fields

creationTime :: Maybe POSIX

When the pipeline was created.

arn :: Maybe Text

The ARN of the pipeline.

activities :: Maybe (NonEmpty PipelineActivity)

The activities that perform transformations on the messages.

name :: Maybe Text

The name of the pipeline.

reprocessingSummaries :: Maybe [ReprocessingSummary]

A summary of information about the pipeline reprocessing.

lastUpdateTime :: Maybe POSIX

The last time the pipeline was updated.

newPipeline :: Pipeline

Create a value of Pipeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Pipeline', pipeline_creationTime - When the pipeline was created.

$sel:arn:Pipeline', pipeline_arn - The ARN of the pipeline.

$sel:activities:Pipeline', pipeline_activities - The activities that perform transformations on the messages.

$sel:name:Pipeline', pipeline_name - The name of the pipeline.

$sel:reprocessingSummaries:Pipeline', pipeline_reprocessingSummaries - A summary of information about the pipeline reprocessing.

$sel:lastUpdateTime:Pipeline', pipeline_lastUpdateTime - The last time the pipeline was updated.

pipeline_creationTime :: Lens' Pipeline (Maybe UTCTime)

When the pipeline was created.

pipeline_arn :: Lens' Pipeline (Maybe Text)

The ARN of the pipeline.

pipeline_activities :: Lens' Pipeline (Maybe (NonEmpty PipelineActivity))

The activities that perform transformations on the messages.

pipeline_name :: Lens' Pipeline (Maybe Text)

The name of the pipeline.

pipeline_reprocessingSummaries :: Lens' Pipeline (Maybe [ReprocessingSummary])

A summary of information about the pipeline reprocessing.

pipeline_lastUpdateTime :: Lens' Pipeline (Maybe UTCTime)

The last time the pipeline was updated.