Network.AWS.IoTAnalytics.Types.SelectAttributesActivity

Description

Documentation

data SelectAttributesActivity

Used to create a new message using only the specified attributes from the original message.

See: newSelectAttributesActivity smart constructor.

Constructors

SelectAttributesActivity'

Fields

next :: Maybe Text

The next activity in the pipeline.

name :: Text

The name of the selectAttributes activity.

attributes :: NonEmpty Text

A list of the attributes to select from the message.

newSelectAttributesActivity

Arguments

:: Text
$sel:name:SelectAttributesActivity'

-> NonEmpty Text
$sel:attributes:SelectAttributesActivity'

-> SelectAttributesActivity

Create a value of SelectAttributesActivity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:next:SelectAttributesActivity', selectAttributesActivity_next - The next activity in the pipeline.

$sel:name:SelectAttributesActivity', selectAttributesActivity_name - The name of the selectAttributes activity.

$sel:attributes:SelectAttributesActivity', selectAttributesActivity_attributes - A list of the attributes to select from the message.

selectAttributesActivity_next :: Lens' SelectAttributesActivity (Maybe Text)

The next activity in the pipeline.

selectAttributesActivity_name :: Lens' SelectAttributesActivity Text

The name of the selectAttributes activity.

selectAttributesActivity_attributes :: Lens' SelectAttributesActivity (NonEmpty Text)

A list of the attributes to select from the message.