Network.AWS.IoTAnalytics.Types.VersioningConfiguration

Description

Documentation

data VersioningConfiguration

Information about the versioning of dataset contents.

See: newVersioningConfiguration smart constructor.

Constructors

VersioningConfiguration'

Fields

maxVersions :: Maybe Natural

How many versions of dataset contents are kept. The unlimited parameter must be false.

unlimited :: Maybe Bool

If true, unlimited versions of dataset contents are kept.

newVersioningConfiguration :: VersioningConfiguration

Create a value of VersioningConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxVersions:VersioningConfiguration', versioningConfiguration_maxVersions - How many versions of dataset contents are kept. The unlimited parameter must be false.

$sel:unlimited:VersioningConfiguration', versioningConfiguration_unlimited - If true, unlimited versions of dataset contents are kept.

versioningConfiguration_maxVersions :: Lens' VersioningConfiguration (Maybe Natural)

How many versions of dataset contents are kept. The unlimited parameter must be false.

versioningConfiguration_unlimited :: Lens' VersioningConfiguration (Maybe Bool)

If true, unlimited versions of dataset contents are kept.