Network.AWS.IoTAnalytics.UpdatePipeline

Description

Updates the settings of a pipeline. You must specify both a channel and a datastore activity and, optionally, as many as 23 additional activities in the pipelineActivities array.

Creating a Request

data UpdatePipeline

See: newUpdatePipeline smart constructor.

Constructors

UpdatePipeline'

Fields

pipelineName :: Text

The name of the pipeline to update.

pipelineActivities :: NonEmpty PipelineActivity

A list of PipelineActivity objects. Activities perform transformations on your messages, such as removing, renaming or adding message attributes; filtering messages based on attribute values; invoking your Lambda functions on messages for advanced processing; or performing mathematical transformations to normalize device data.

The list can be 2-25 PipelineActivity objects and must contain both a channel and a datastore activity. Each entry in the list must contain only one activity. For example:

pipelineActivities = [ { "channel": { ... } }, { "lambda": { ... } }, ... ]

newUpdatePipeline

Arguments

:: Text
$sel:pipelineName:UpdatePipeline'

-> NonEmpty PipelineActivity
$sel:pipelineActivities:UpdatePipeline'

-> UpdatePipeline

Create a value of UpdatePipeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pipelineName:UpdatePipeline', updatePipeline_pipelineName - The name of the pipeline to update.

$sel:pipelineActivities:UpdatePipeline', updatePipeline_pipelineActivities - A list of PipelineActivity objects. Activities perform transformations on your messages, such as removing, renaming or adding message attributes; filtering messages based on attribute values; invoking your Lambda functions on messages for advanced processing; or performing mathematical transformations to normalize device data.

The list can be 2-25 PipelineActivity objects and must contain both a channel and a datastore activity. Each entry in the list must contain only one activity. For example:

pipelineActivities = [ { "channel": { ... } }, { "lambda": { ... } }, ... ]

Request Lenses

updatePipeline_pipelineName :: Lens' UpdatePipeline Text

The name of the pipeline to update.

updatePipeline_pipelineActivities :: Lens' UpdatePipeline (NonEmpty PipelineActivity)

A list of PipelineActivity objects. Activities perform transformations on your messages, such as removing, renaming or adding message attributes; filtering messages based on attribute values; invoking your Lambda functions on messages for advanced processing; or performing mathematical transformations to normalize device data.

The list can be 2-25 PipelineActivity objects and must contain both a channel and a datastore activity. Each entry in the list must contain only one activity. For example:

pipelineActivities = [ { "channel": { ... } }, { "lambda": { ... } }, ... ]

Destructuring the Response

data UpdatePipelineResponse

See: newUpdatePipelineResponse smart constructor.

Constructors

UpdatePipelineResponse'

newUpdatePipelineResponse :: UpdatePipelineResponse

Create a value of UpdatePipelineResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.