Network.AWS.IoTData

Description

Derived from API version 2015-05-28 of the AWS service descriptions, licensed under Apache 2.0.

IoT data

IoT data enables secure, bi-directional communication between Internet-connected things (such as sensors, actuators, embedded devices, or smart appliances) and the Amazon Web Services cloud. It implements a broker for applications and things to publish messages over HTTP (Publish) and retrieve, update, and delete shadows. A shadow is a persistent representation of your things and their state in the Amazon Web Services cloud.

Find the endpoint address for actions in IoT data by running this CLI command:

aws iot describe-endpoint --endpoint-type iot:Data-ATS

The service name used by Amazon Web ServicesSignature Version 4 to sign requests is: iotdevicegateway.

Service Configuration

defaultService :: Service

API version 2015-05-28 of the Amazon IoT Data Plane SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by IoTData.

UnsupportedDocumentEncodingException

_UnsupportedDocumentEncodingException :: AsError a => Getting (First ServiceError) a ServiceError

The document encoding is not supported.

UnauthorizedException

_UnauthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

You are not authorized to perform this operation.

ServiceUnavailableException

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The service is temporarily unavailable.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The rate exceeds the limit.

InvalidRequestException

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request is not valid.

RequestEntityTooLargeException

_RequestEntityTooLargeException :: AsError a => Getting (First ServiceError) a ServiceError

The payload exceeds the maximum size allowed.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The specified version does not match the version of the document.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource does not exist.

InternalFailureException

_InternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

An unexpected error has occurred.

MethodNotAllowedException

_MethodNotAllowedException :: AsError a => Getting (First ServiceError) a ServiceError

The specified combination of HTTP verb and URI is not supported.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

GetThingShadow

data GetThingShadow

The input for the GetThingShadow operation.

See: newGetThingShadow smart constructor.

Constructors

GetThingShadow' (Maybe Text) Text

newGetThingShadow

Arguments

:: Text
$sel:thingName:GetThingShadow'

-> GetThingShadow

Create a value of GetThingShadow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shadowName:GetThingShadow', getThingShadow_shadowName - The name of the shadow.

$sel:thingName:GetThingShadow', getThingShadow_thingName - The name of the thing.

data GetThingShadowResponse

The output from the GetThingShadow operation.

See: newGetThingShadowResponse smart constructor.

Constructors

GetThingShadowResponse' (Maybe ByteString) Int

newGetThingShadowResponse

Arguments

:: Int
$sel:httpStatus:GetThingShadowResponse'

-> GetThingShadowResponse

Create a value of GetThingShadowResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:payload:GetThingShadowResponse', getThingShadowResponse_payload - The state information, in JSON format.

$sel:httpStatus:GetThingShadowResponse', getThingShadowResponse_httpStatus - The response's http status code.

Publish

data Publish

The input for the Publish operation.

See: newPublish smart constructor.

Constructors

Publish' (Maybe ByteString) (Maybe Bool) (Maybe Natural) Text

newPublish

Arguments

:: Text
$sel:topic:Publish'

-> Publish

Create a value of Publish with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:payload:Publish', publish_payload - The message body. MQTT accepts text, binary, and empty (null) message payloads.

Publishing an empty (null) payload with retain = true deletes the retained message identified by topic from IoT Core.

$sel:retain:Publish', publish_retain - A Boolean value that determines whether to set the RETAIN flag when the message is published.

Setting the RETAIN flag causes the message to be retained and sent to new subscribers to the topic.

Valid values: true | false

Default value: false

$sel:qos:Publish', publish_qos - The Quality of Service (QoS) level.

$sel:topic:Publish', publish_topic - The name of the MQTT topic.

data PublishResponse

See: newPublishResponse smart constructor.

Constructors

PublishResponse'

newPublishResponse :: PublishResponse

Create a value of PublishResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

ListRetainedMessages (Paginated)

data ListRetainedMessages

See: newListRetainedMessages smart constructor.

Constructors

ListRetainedMessages' (Maybe Text) (Maybe Natural)

newListRetainedMessages :: ListRetainedMessages

Create a value of ListRetainedMessages with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRetainedMessages', listRetainedMessages_nextToken - To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

$sel:maxResults:ListRetainedMessages', listRetainedMessages_maxResults - The maximum number of results to return at one time.

data ListRetainedMessagesResponse

See: newListRetainedMessagesResponse smart constructor.

Constructors

ListRetainedMessagesResponse' (Maybe Text) (Maybe [RetainedMessageSummary]) Int

newListRetainedMessagesResponse

Arguments

:: Int
$sel:httpStatus:ListRetainedMessagesResponse'

-> ListRetainedMessagesResponse

Create a value of ListRetainedMessagesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListRetainedMessages', listRetainedMessagesResponse_nextToken - The token for the next set of results, or null if there are no additional results.

$sel:retainedTopics:ListRetainedMessagesResponse', listRetainedMessagesResponse_retainedTopics - A summary list the account's retained messages. The information returned doesn't include the message payloads of the retained messages.

$sel:httpStatus:ListRetainedMessagesResponse', listRetainedMessagesResponse_httpStatus - The response's http status code.

ListNamedShadowsForThing

data ListNamedShadowsForThing

See: newListNamedShadowsForThing smart constructor.

Constructors

ListNamedShadowsForThing' (Maybe Text) (Maybe Natural) Text

newListNamedShadowsForThing

Arguments

:: Text
$sel:thingName:ListNamedShadowsForThing'

-> ListNamedShadowsForThing

Create a value of ListNamedShadowsForThing with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListNamedShadowsForThing', listNamedShadowsForThing_nextToken - The token to retrieve the next set of results.

$sel:pageSize:ListNamedShadowsForThing', listNamedShadowsForThing_pageSize - The result page size.

$sel:thingName:ListNamedShadowsForThing', listNamedShadowsForThing_thingName - The name of the thing.

data ListNamedShadowsForThingResponse

See: newListNamedShadowsForThingResponse smart constructor.

Constructors

ListNamedShadowsForThingResponse' (Maybe Text) (Maybe Integer) (Maybe [Text]) Int

newListNamedShadowsForThingResponse

Arguments

:: Int
$sel:httpStatus:ListNamedShadowsForThingResponse'

-> ListNamedShadowsForThingResponse

Create a value of ListNamedShadowsForThingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListNamedShadowsForThing', listNamedShadowsForThingResponse_nextToken - The token to use to get the next set of results, or null if there are no additional results.

$sel:timestamp:ListNamedShadowsForThingResponse', listNamedShadowsForThingResponse_timestamp - The Epoch date and time the response was generated by IoT.

$sel:results:ListNamedShadowsForThingResponse', listNamedShadowsForThingResponse_results - The list of shadows for the specified thing.

$sel:httpStatus:ListNamedShadowsForThingResponse', listNamedShadowsForThingResponse_httpStatus - The response's http status code.

UpdateThingShadow

data UpdateThingShadow

The input for the UpdateThingShadow operation.

See: newUpdateThingShadow smart constructor.

Constructors

UpdateThingShadow' (Maybe Text) Text ByteString

newUpdateThingShadow

Arguments

:: Text
$sel:thingName:UpdateThingShadow'

-> ByteString
$sel:payload:UpdateThingShadow'

-> UpdateThingShadow

Create a value of UpdateThingShadow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shadowName:UpdateThingShadow', updateThingShadow_shadowName - The name of the shadow.

$sel:thingName:UpdateThingShadow', updateThingShadow_thingName - The name of the thing.

$sel:payload:UpdateThingShadow', updateThingShadow_payload - The state information, in JSON format.

data UpdateThingShadowResponse

The output from the UpdateThingShadow operation.

See: newUpdateThingShadowResponse smart constructor.

Constructors

UpdateThingShadowResponse' (Maybe ByteString) Int

newUpdateThingShadowResponse

Arguments

:: Int
$sel:httpStatus:UpdateThingShadowResponse'

-> UpdateThingShadowResponse

Create a value of UpdateThingShadowResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:payload:UpdateThingShadow', updateThingShadowResponse_payload - The state information, in JSON format.

$sel:httpStatus:UpdateThingShadowResponse', updateThingShadowResponse_httpStatus - The response's http status code.

DeleteThingShadow

data DeleteThingShadow

The input for the DeleteThingShadow operation.

See: newDeleteThingShadow smart constructor.

Constructors

DeleteThingShadow' (Maybe Text) Text

newDeleteThingShadow

Arguments

:: Text
$sel:thingName:DeleteThingShadow'

-> DeleteThingShadow

Create a value of DeleteThingShadow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shadowName:DeleteThingShadow', deleteThingShadow_shadowName - The name of the shadow.

$sel:thingName:DeleteThingShadow', deleteThingShadow_thingName - The name of the thing.

data DeleteThingShadowResponse

The output from the DeleteThingShadow operation.

See: newDeleteThingShadowResponse smart constructor.

Constructors

DeleteThingShadowResponse' Int ByteString

newDeleteThingShadowResponse

Arguments

:: Int
$sel:httpStatus:DeleteThingShadowResponse'

-> ByteString
$sel:payload:DeleteThingShadowResponse'

-> DeleteThingShadowResponse

Create a value of DeleteThingShadowResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteThingShadowResponse', deleteThingShadowResponse_httpStatus - The response's http status code.

$sel:payload:DeleteThingShadowResponse', deleteThingShadowResponse_payload - The state information, in JSON format.

GetRetainedMessage

data GetRetainedMessage

The input for the GetRetainedMessage operation.

See: newGetRetainedMessage smart constructor.

Constructors

GetRetainedMessage' Text

newGetRetainedMessage

Arguments

:: Text
$sel:topic:GetRetainedMessage'

-> GetRetainedMessage

Create a value of GetRetainedMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:topic:GetRetainedMessage', getRetainedMessage_topic - The topic name of the retained message to retrieve.

data GetRetainedMessageResponse

The output from the GetRetainedMessage operation.

See: newGetRetainedMessageResponse smart constructor.

Constructors

GetRetainedMessageResponse' (Maybe Base64) (Maybe Text) (Maybe Integer) (Maybe Natural) Int

newGetRetainedMessageResponse

Arguments

:: Int
$sel:httpStatus:GetRetainedMessageResponse'

-> GetRetainedMessageResponse

Create a value of GetRetainedMessageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:payload:GetRetainedMessageResponse', getRetainedMessageResponse_payload - The Base64-encoded message payload of the retained message body.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:topic:GetRetainedMessage', getRetainedMessageResponse_topic - The topic name to which the retained message was published.

$sel:lastModifiedTime:GetRetainedMessageResponse', getRetainedMessageResponse_lastModifiedTime - The Epoch date and time, in milliseconds, when the retained message was stored by IoT.

$sel:qos:GetRetainedMessageResponse', getRetainedMessageResponse_qos - The quality of service (QoS) level used to publish the retained message.

$sel:httpStatus:GetRetainedMessageResponse', getRetainedMessageResponse_httpStatus - The response's http status code.

Types

RetainedMessageSummary

data RetainedMessageSummary

Information about a single retained message.

See: newRetainedMessageSummary smart constructor.

Constructors

RetainedMessageSummary' (Maybe Integer) (Maybe Text) (Maybe Integer) (Maybe Natural)

newRetainedMessageSummary :: RetainedMessageSummary

Create a value of RetainedMessageSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:payloadSize:RetainedMessageSummary', retainedMessageSummary_payloadSize - The size of the retained message's payload in bytes.

$sel:topic:RetainedMessageSummary', retainedMessageSummary_topic - The topic name to which the retained message was published.

$sel:lastModifiedTime:RetainedMessageSummary', retainedMessageSummary_lastModifiedTime - The Epoch date and time, in milliseconds, when the retained message was stored by IoT.

$sel:qos:RetainedMessageSummary', retainedMessageSummary_qos - The quality of service (QoS) level used to publish the retained message.