Network.AWS.IoTData.Lens

Description

Operations

GetThingShadow

getThingShadow_shadowName :: Lens' GetThingShadow (Maybe Text)

The name of the shadow.

getThingShadow_thingName :: Lens' GetThingShadow Text

The name of the thing.

getThingShadowResponse_payload :: Lens' GetThingShadowResponse (Maybe ByteString)

The state information, in JSON format.

getThingShadowResponse_httpStatus :: Lens' GetThingShadowResponse Int

The response's http status code.

ListNamedShadowsForThing

listNamedShadowsForThing_nextToken :: Lens' ListNamedShadowsForThing (Maybe Text)

The token to retrieve the next set of results.

listNamedShadowsForThing_pageSize :: Lens' ListNamedShadowsForThing (Maybe Natural)

The result page size.

listNamedShadowsForThing_thingName :: Lens' ListNamedShadowsForThing Text

The name of the thing.

listNamedShadowsForThingResponse_results :: Lens' ListNamedShadowsForThingResponse (Maybe [Text])

The list of shadows for the specified thing.

listNamedShadowsForThingResponse_nextToken :: Lens' ListNamedShadowsForThingResponse (Maybe Text)

The token to use to get the next set of results, or null if there are no additional results.

listNamedShadowsForThingResponse_timestamp :: Lens' ListNamedShadowsForThingResponse (Maybe Integer)

The Epoch date and time the response was generated by IoT.

listNamedShadowsForThingResponse_httpStatus :: Lens' ListNamedShadowsForThingResponse Int

The response's http status code.

DeleteThingShadow

deleteThingShadow_shadowName :: Lens' DeleteThingShadow (Maybe Text)

The name of the shadow.

deleteThingShadow_thingName :: Lens' DeleteThingShadow Text

The name of the thing.

deleteThingShadowResponse_httpStatus :: Lens' DeleteThingShadowResponse Int

The response's http status code.

deleteThingShadowResponse_payload :: Lens' DeleteThingShadowResponse ByteString

The state information, in JSON format.

UpdateThingShadow

updateThingShadow_shadowName :: Lens' UpdateThingShadow (Maybe Text)

The name of the shadow.

updateThingShadow_thingName :: Lens' UpdateThingShadow Text

The name of the thing.

updateThingShadow_payload :: Lens' UpdateThingShadow ByteString

The state information, in JSON format.

updateThingShadowResponse_payload :: Lens' UpdateThingShadowResponse (Maybe ByteString)

The state information, in JSON format.

updateThingShadowResponse_httpStatus :: Lens' UpdateThingShadowResponse Int

The response's http status code.

ListRetainedMessages

listRetainedMessages_nextToken :: Lens' ListRetainedMessages (Maybe Text)

To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

listRetainedMessages_maxResults :: Lens' ListRetainedMessages (Maybe Natural)

The maximum number of results to return at one time.

listRetainedMessagesResponse_retainedTopics :: Lens' ListRetainedMessagesResponse (Maybe [RetainedMessageSummary])

A summary list the account's retained messages. The information returned doesn't include the message payloads of the retained messages.

listRetainedMessagesResponse_nextToken :: Lens' ListRetainedMessagesResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

listRetainedMessagesResponse_httpStatus :: Lens' ListRetainedMessagesResponse Int

The response's http status code.

GetRetainedMessage

getRetainedMessage_topic :: Lens' GetRetainedMessage Text

The topic name of the retained message to retrieve.

getRetainedMessageResponse_lastModifiedTime :: Lens' GetRetainedMessageResponse (Maybe Integer)

The Epoch date and time, in milliseconds, when the retained message was stored by IoT.

getRetainedMessageResponse_payload :: Lens' GetRetainedMessageResponse (Maybe ByteString)

The Base64-encoded message payload of the retained message body.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

getRetainedMessageResponse_topic :: Lens' GetRetainedMessageResponse (Maybe Text)

The topic name to which the retained message was published.

getRetainedMessageResponse_qos :: Lens' GetRetainedMessageResponse (Maybe Natural)

The quality of service (QoS) level used to publish the retained message.

getRetainedMessageResponse_httpStatus :: Lens' GetRetainedMessageResponse Int

The response's http status code.

Publish

publish_retain :: Lens' Publish (Maybe Bool)

A Boolean value that determines whether to set the RETAIN flag when the message is published.

Setting the RETAIN flag causes the message to be retained and sent to new subscribers to the topic.

Valid values: true | false

Default value: false

publish_payload :: Lens' Publish (Maybe ByteString)

The message body. MQTT accepts text, binary, and empty (null) message payloads.

Publishing an empty (null) payload with retain = true deletes the retained message identified by topic from IoT Core.

publish_qos :: Lens' Publish (Maybe Natural)

The Quality of Service (QoS) level.

publish_topic :: Lens' Publish Text

The name of the MQTT topic.

Types

RetainedMessageSummary

retainedMessageSummary_lastModifiedTime :: Lens' RetainedMessageSummary (Maybe Integer)

The Epoch date and time, in milliseconds, when the retained message was stored by IoT.

retainedMessageSummary_topic :: Lens' RetainedMessageSummary (Maybe Text)

The topic name to which the retained message was published.

retainedMessageSummary_qos :: Lens' RetainedMessageSummary (Maybe Natural)

The quality of service (QoS) level used to publish the retained message.

retainedMessageSummary_payloadSize :: Lens' RetainedMessageSummary (Maybe Integer)

The size of the retained message's payload in bytes.