Network.AWS.IoTData.ListNamedShadowsForThing

Description

Lists the shadows for the specified thing.

Requires permission to access the ListNamedShadowsForThing action.

Creating a Request

data ListNamedShadowsForThing

See: newListNamedShadowsForThing smart constructor.

Constructors

ListNamedShadowsForThing'

Fields

nextToken :: Maybe Text

The token to retrieve the next set of results.

pageSize :: Maybe Natural

The result page size.

thingName :: Text

The name of the thing.

newListNamedShadowsForThing

Arguments

:: Text
$sel:thingName:ListNamedShadowsForThing'

-> ListNamedShadowsForThing

Create a value of ListNamedShadowsForThing with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListNamedShadowsForThing', listNamedShadowsForThing_nextToken - The token to retrieve the next set of results.

$sel:pageSize:ListNamedShadowsForThing', listNamedShadowsForThing_pageSize - The result page size.

$sel:thingName:ListNamedShadowsForThing', listNamedShadowsForThing_thingName - The name of the thing.

Request Lenses

listNamedShadowsForThing_nextToken :: Lens' ListNamedShadowsForThing (Maybe Text)

The token to retrieve the next set of results.

listNamedShadowsForThing_pageSize :: Lens' ListNamedShadowsForThing (Maybe Natural)

The result page size.

listNamedShadowsForThing_thingName :: Lens' ListNamedShadowsForThing Text

The name of the thing.

Destructuring the Response

data ListNamedShadowsForThingResponse

See: newListNamedShadowsForThingResponse smart constructor.

Constructors

ListNamedShadowsForThingResponse'

Fields

results :: Maybe [Text]

The list of shadows for the specified thing.

nextToken :: Maybe Text

The token to use to get the next set of results, or null if there are no additional results.

timestamp :: Maybe Integer

The Epoch date and time the response was generated by IoT.

httpStatus :: Int

The response's http status code.

newListNamedShadowsForThingResponse

Arguments

:: Int
$sel:httpStatus:ListNamedShadowsForThingResponse'

-> ListNamedShadowsForThingResponse

Create a value of ListNamedShadowsForThingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:results:ListNamedShadowsForThingResponse', listNamedShadowsForThingResponse_results - The list of shadows for the specified thing.

$sel:nextToken:ListNamedShadowsForThing', listNamedShadowsForThingResponse_nextToken - The token to use to get the next set of results, or null if there are no additional results.

$sel:timestamp:ListNamedShadowsForThingResponse', listNamedShadowsForThingResponse_timestamp - The Epoch date and time the response was generated by IoT.

$sel:httpStatus:ListNamedShadowsForThingResponse', listNamedShadowsForThingResponse_httpStatus - The response's http status code.

Response Lenses

listNamedShadowsForThingResponse_results :: Lens' ListNamedShadowsForThingResponse (Maybe [Text])

The list of shadows for the specified thing.

listNamedShadowsForThingResponse_nextToken :: Lens' ListNamedShadowsForThingResponse (Maybe Text)

The token to use to get the next set of results, or null if there are no additional results.

listNamedShadowsForThingResponse_timestamp :: Lens' ListNamedShadowsForThingResponse (Maybe Integer)

The Epoch date and time the response was generated by IoT.

listNamedShadowsForThingResponse_httpStatus :: Lens' ListNamedShadowsForThingResponse Int

The response's http status code.