Network.AWS.IoTData.Publish

Description

Publishes an MQTT message.

Requires permission to access the Publish action.

For more information about MQTT messages, see MQTT Protocol in the IoT Developer Guide.

For more information about messaging costs, see IoT Core pricing - Messaging.

Creating a Request

data Publish

The input for the Publish operation.

See: newPublish smart constructor.

Constructors

Publish'

Fields

payload :: Maybe ByteString

The message body. MQTT accepts text, binary, and empty (null) message payloads.

Publishing an empty (null) payload with retain = true deletes the retained message identified by topic from IoT Core.

retain :: Maybe Bool

A Boolean value that determines whether to set the RETAIN flag when the message is published.

Setting the RETAIN flag causes the message to be retained and sent to new subscribers to the topic.

Valid values: true | false

Default value: false

qos :: Maybe Natural

The Quality of Service (QoS) level.

topic :: Text

The name of the MQTT topic.

newPublish

Arguments

:: Text
$sel:topic:Publish'

-> Publish

Create a value of Publish with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:payload:Publish', publish_payload - The message body. MQTT accepts text, binary, and empty (null) message payloads.

Publishing an empty (null) payload with retain = true deletes the retained message identified by topic from IoT Core.

$sel:retain:Publish', publish_retain - A Boolean value that determines whether to set the RETAIN flag when the message is published.

Setting the RETAIN flag causes the message to be retained and sent to new subscribers to the topic.

Valid values: true | false

Default value: false

$sel:qos:Publish', publish_qos - The Quality of Service (QoS) level.

$sel:topic:Publish', publish_topic - The name of the MQTT topic.

Request Lenses

publish_payload :: Lens' Publish (Maybe ByteString)

The message body. MQTT accepts text, binary, and empty (null) message payloads.

Publishing an empty (null) payload with retain = true deletes the retained message identified by topic from IoT Core.

publish_retain :: Lens' Publish (Maybe Bool)

A Boolean value that determines whether to set the RETAIN flag when the message is published.

Setting the RETAIN flag causes the message to be retained and sent to new subscribers to the topic.

Valid values: true | false

Default value: false

publish_qos :: Lens' Publish (Maybe Natural)

The Quality of Service (QoS) level.

publish_topic :: Lens' Publish Text

The name of the MQTT topic.

Destructuring the Response

data PublishResponse

See: newPublishResponse smart constructor.

Constructors

PublishResponse'

newPublishResponse :: PublishResponse

Create a value of PublishResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.