Network.AWS.IoTData.Types

Description

Service Configuration

defaultService :: Service

API version 2015-05-28 of the Amazon IoT Data Plane SDK configuration.

Errors

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request is not valid.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The specified version does not match the version of the document.

_RequestEntityTooLargeException :: AsError a => Getting (First ServiceError) a ServiceError

The payload exceeds the maximum size allowed.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The rate exceeds the limit.

_MethodNotAllowedException :: AsError a => Getting (First ServiceError) a ServiceError

The specified combination of HTTP verb and URI is not supported.

_InternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

An unexpected error has occurred.

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The service is temporarily unavailable.

_UnauthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

You are not authorized to perform this operation.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource does not exist.

_UnsupportedDocumentEncodingException :: AsError a => Getting (First ServiceError) a ServiceError

The document encoding is not supported.

RetainedMessageSummary

data RetainedMessageSummary

Information about a single retained message.

See: newRetainedMessageSummary smart constructor.

Constructors

RetainedMessageSummary'

Fields

lastModifiedTime :: Maybe Integer

The Epoch date and time, in milliseconds, when the retained message was stored by IoT.

topic :: Maybe Text

The topic name to which the retained message was published.

qos :: Maybe Natural

The quality of service (QoS) level used to publish the retained message.

payloadSize :: Maybe Integer

The size of the retained message's payload in bytes.

newRetainedMessageSummary :: RetainedMessageSummary

Create a value of RetainedMessageSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedTime:RetainedMessageSummary', retainedMessageSummary_lastModifiedTime - The Epoch date and time, in milliseconds, when the retained message was stored by IoT.

$sel:topic:RetainedMessageSummary', retainedMessageSummary_topic - The topic name to which the retained message was published.

$sel:qos:RetainedMessageSummary', retainedMessageSummary_qos - The quality of service (QoS) level used to publish the retained message.

$sel:payloadSize:RetainedMessageSummary', retainedMessageSummary_payloadSize - The size of the retained message's payload in bytes.

retainedMessageSummary_lastModifiedTime :: Lens' RetainedMessageSummary (Maybe Integer)

The Epoch date and time, in milliseconds, when the retained message was stored by IoT.

retainedMessageSummary_topic :: Lens' RetainedMessageSummary (Maybe Text)

The topic name to which the retained message was published.

retainedMessageSummary_qos :: Lens' RetainedMessageSummary (Maybe Natural)

The quality of service (QoS) level used to publish the retained message.

retainedMessageSummary_payloadSize :: Lens' RetainedMessageSummary (Maybe Integer)

The size of the retained message's payload in bytes.