Network.AWS.IoTDeviceAdvisor.CreateSuiteDefinition

Description

Creates a Device Advisor test suite.

Creating a Request

data CreateSuiteDefinition

See: newCreateSuiteDefinition smart constructor.

Constructors

CreateSuiteDefinition'

Fields

suiteDefinitionConfiguration :: Maybe SuiteDefinitionConfiguration

Creates a Device Advisor test suite with suite definition configuration.

tags :: Maybe (HashMap Text Text)

The tags to be attached to the suite definition.

newCreateSuiteDefinition :: CreateSuiteDefinition

Create a value of CreateSuiteDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suiteDefinitionConfiguration:CreateSuiteDefinition', createSuiteDefinition_suiteDefinitionConfiguration - Creates a Device Advisor test suite with suite definition configuration.

$sel:tags:CreateSuiteDefinition', createSuiteDefinition_tags - The tags to be attached to the suite definition.

Request Lenses

createSuiteDefinition_suiteDefinitionConfiguration :: Lens' CreateSuiteDefinition (Maybe SuiteDefinitionConfiguration)

Creates a Device Advisor test suite with suite definition configuration.

createSuiteDefinition_tags :: Lens' CreateSuiteDefinition (Maybe (HashMap Text Text))

The tags to be attached to the suite definition.

Destructuring the Response

data CreateSuiteDefinitionResponse

See: newCreateSuiteDefinitionResponse smart constructor.

Constructors

CreateSuiteDefinitionResponse'

Fields

createdAt :: Maybe POSIX

Creates a Device Advisor test suite with TimeStamp of when it was created.

suiteDefinitionArn :: Maybe Text

Creates a Device Advisor test suite with Amazon Resource name.

suiteDefinitionId :: Maybe Text

Creates a Device Advisor test suite with suite UUID.

suiteDefinitionName :: Maybe Text

Creates a Device Advisor test suite with suite definition name.

httpStatus :: Int

The response's http status code.

newCreateSuiteDefinitionResponse

Arguments

:: Int
$sel:httpStatus:CreateSuiteDefinitionResponse'

-> CreateSuiteDefinitionResponse

Create a value of CreateSuiteDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAt:CreateSuiteDefinitionResponse', createSuiteDefinitionResponse_createdAt - Creates a Device Advisor test suite with TimeStamp of when it was created.

$sel:suiteDefinitionArn:CreateSuiteDefinitionResponse', createSuiteDefinitionResponse_suiteDefinitionArn - Creates a Device Advisor test suite with Amazon Resource name.

$sel:suiteDefinitionId:CreateSuiteDefinitionResponse', createSuiteDefinitionResponse_suiteDefinitionId - Creates a Device Advisor test suite with suite UUID.

$sel:suiteDefinitionName:CreateSuiteDefinitionResponse', createSuiteDefinitionResponse_suiteDefinitionName - Creates a Device Advisor test suite with suite definition name.

$sel:httpStatus:CreateSuiteDefinitionResponse', createSuiteDefinitionResponse_httpStatus - The response's http status code.

Response Lenses

createSuiteDefinitionResponse_createdAt :: Lens' CreateSuiteDefinitionResponse (Maybe UTCTime)

Creates a Device Advisor test suite with TimeStamp of when it was created.

createSuiteDefinitionResponse_suiteDefinitionArn :: Lens' CreateSuiteDefinitionResponse (Maybe Text)

Creates a Device Advisor test suite with Amazon Resource name.

createSuiteDefinitionResponse_suiteDefinitionId :: Lens' CreateSuiteDefinitionResponse (Maybe Text)

Creates a Device Advisor test suite with suite UUID.

createSuiteDefinitionResponse_suiteDefinitionName :: Lens' CreateSuiteDefinitionResponse (Maybe Text)

Creates a Device Advisor test suite with suite definition name.

createSuiteDefinitionResponse_httpStatus :: Lens' CreateSuiteDefinitionResponse Int

The response's http status code.