Network.AWS.IoTDeviceAdvisor.Lens

Description

Operations

GetSuiteRunReport

getSuiteRunReport_suiteDefinitionId :: Lens' GetSuiteRunReport Text

Suite definition Id of the test suite.

getSuiteRunReport_suiteRunId :: Lens' GetSuiteRunReport Text

Suite run Id of the test suite run.

getSuiteRunReportResponse_qualificationReportDownloadUrl :: Lens' GetSuiteRunReportResponse (Maybe Text)

Download URL of the qualification report.

getSuiteRunReportResponse_httpStatus :: Lens' GetSuiteRunReportResponse Int

The response's http status code.

StartSuiteRun

startSuiteRun_suiteRunConfiguration :: Lens' StartSuiteRun (Maybe SuiteRunConfiguration)

Suite run configuration.

startSuiteRun_suiteDefinitionVersion :: Lens' StartSuiteRun (Maybe Text)

Suite definition version of the test suite.

startSuiteRun_tags :: Lens' StartSuiteRun (Maybe (HashMap Text Text))

The tags to be attached to the suite run.

startSuiteRun_suiteDefinitionId :: Lens' StartSuiteRun Text

Suite definition Id of the test suite.

startSuiteRunResponse_createdAt :: Lens' StartSuiteRunResponse (Maybe UTCTime)

Date (in Unix epoch time) when the suite run was created.

startSuiteRunResponse_suiteRunArn :: Lens' StartSuiteRunResponse (Maybe Text)

Amazon resource name of the started suite run.

startSuiteRunResponse_suiteRunId :: Lens' StartSuiteRunResponse (Maybe Text)

Suite Run Id of the started suite run.

startSuiteRunResponse_httpStatus :: Lens' StartSuiteRunResponse Int

The response's http status code.

ListSuiteDefinitions

listSuiteDefinitions_nextToken :: Lens' ListSuiteDefinitions (Maybe Text)

A token used to get the next set of results.

listSuiteDefinitions_maxResults :: Lens' ListSuiteDefinitions (Maybe Natural)

The maximum number of results to return at once.

listSuiteDefinitionsResponse_suiteDefinitionInformationList :: Lens' ListSuiteDefinitionsResponse (Maybe [SuiteDefinitionInformation])

An array of objects that provide summaries of information about the suite definitions in the list.

listSuiteDefinitionsResponse_nextToken :: Lens' ListSuiteDefinitionsResponse (Maybe Text)

A token used to get the next set of results.

listSuiteDefinitionsResponse_httpStatus :: Lens' ListSuiteDefinitionsResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The ARN of the IoT Device Advisor resource.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The tags attached to the IoT Device Advisor resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

DeleteSuiteDefinition

deleteSuiteDefinition_suiteDefinitionId :: Lens' DeleteSuiteDefinition Text

Suite definition Id of the test suite to be deleted.

deleteSuiteDefinitionResponse_httpStatus :: Lens' DeleteSuiteDefinitionResponse Int

The response's http status code.

UpdateSuiteDefinition

updateSuiteDefinition_suiteDefinitionConfiguration :: Lens' UpdateSuiteDefinition (Maybe SuiteDefinitionConfiguration)

Updates a Device Advisor test suite with suite definition configuration.

updateSuiteDefinition_suiteDefinitionId :: Lens' UpdateSuiteDefinition Text

Suite definition Id of the test suite to be updated.

updateSuiteDefinitionResponse_lastUpdatedAt :: Lens' UpdateSuiteDefinitionResponse (Maybe UTCTime)

Timestamp of when the test suite was updated.

updateSuiteDefinitionResponse_createdAt :: Lens' UpdateSuiteDefinitionResponse (Maybe UTCTime)

Timestamp of when the test suite was created.

updateSuiteDefinitionResponse_suiteDefinitionArn :: Lens' UpdateSuiteDefinitionResponse (Maybe Text)

Amazon Resource name of the updated test suite.

updateSuiteDefinitionResponse_suiteDefinitionId :: Lens' UpdateSuiteDefinitionResponse (Maybe Text)

Suite definition Id of the updated test suite.

updateSuiteDefinitionResponse_suiteDefinitionVersion :: Lens' UpdateSuiteDefinitionResponse (Maybe Text)

Suite definition version of the updated test suite.

updateSuiteDefinitionResponse_suiteDefinitionName :: Lens' UpdateSuiteDefinitionResponse (Maybe Text)

Suite definition name of the updated test suite.

updateSuiteDefinitionResponse_httpStatus :: Lens' UpdateSuiteDefinitionResponse Int

The response's http status code.

CreateSuiteDefinition

createSuiteDefinition_suiteDefinitionConfiguration :: Lens' CreateSuiteDefinition (Maybe SuiteDefinitionConfiguration)

Creates a Device Advisor test suite with suite definition configuration.

createSuiteDefinition_tags :: Lens' CreateSuiteDefinition (Maybe (HashMap Text Text))

The tags to be attached to the suite definition.

createSuiteDefinitionResponse_createdAt :: Lens' CreateSuiteDefinitionResponse (Maybe UTCTime)

Creates a Device Advisor test suite with TimeStamp of when it was created.

createSuiteDefinitionResponse_suiteDefinitionArn :: Lens' CreateSuiteDefinitionResponse (Maybe Text)

Creates a Device Advisor test suite with Amazon Resource name.

createSuiteDefinitionResponse_suiteDefinitionId :: Lens' CreateSuiteDefinitionResponse (Maybe Text)

Creates a Device Advisor test suite with suite UUID.

createSuiteDefinitionResponse_suiteDefinitionName :: Lens' CreateSuiteDefinitionResponse (Maybe Text)

Creates a Device Advisor test suite with suite definition name.

createSuiteDefinitionResponse_httpStatus :: Lens' CreateSuiteDefinitionResponse Int

The response's http status code.

StopSuiteRun

stopSuiteRun_suiteDefinitionId :: Lens' StopSuiteRun Text

Suite definition Id of the test suite run to be stopped.

stopSuiteRun_suiteRunId :: Lens' StopSuiteRun Text

Suite run Id of the test suite run to be stopped.

stopSuiteRunResponse_httpStatus :: Lens' StopSuiteRunResponse Int

The response's http status code.

GetSuiteDefinition

getSuiteDefinition_suiteDefinitionVersion :: Lens' GetSuiteDefinition (Maybe Text)

Suite definition version of the test suite to get.

getSuiteDefinition_suiteDefinitionId :: Lens' GetSuiteDefinition Text

Suite definition Id of the test suite to get.

getSuiteDefinitionResponse_createdAt :: Lens' GetSuiteDefinitionResponse (Maybe UTCTime)

Date (in Unix epoch time) when the suite definition was created.

getSuiteDefinitionResponse_suiteDefinitionConfiguration :: Lens' GetSuiteDefinitionResponse (Maybe SuiteDefinitionConfiguration)

Suite configuration of the suite definition.

getSuiteDefinitionResponse_suiteDefinitionArn :: Lens' GetSuiteDefinitionResponse (Maybe Text)

The ARN of the suite definition.

getSuiteDefinitionResponse_lastModifiedAt :: Lens' GetSuiteDefinitionResponse (Maybe UTCTime)

Date (in Unix epoch time) when the suite definition was last modified.

getSuiteDefinitionResponse_suiteDefinitionId :: Lens' GetSuiteDefinitionResponse (Maybe Text)

Suite definition Id of the suite definition.

getSuiteDefinitionResponse_suiteDefinitionVersion :: Lens' GetSuiteDefinitionResponse (Maybe Text)

Suite definition version of the suite definition.

getSuiteDefinitionResponse_latestVersion :: Lens' GetSuiteDefinitionResponse (Maybe Text)

Latest suite definition version of the suite definition.

getSuiteDefinitionResponse_tags :: Lens' GetSuiteDefinitionResponse (Maybe (HashMap Text Text))

Tags attached to the suite definition.

getSuiteDefinitionResponse_httpStatus :: Lens' GetSuiteDefinitionResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The resource ARN of an IoT Device Advisor resource.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The tags to be attached to the IoT Device Advisor resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The resource ARN of an IoT Device Advisor resource.

untagResource_tagKeys :: Lens' UntagResource [Text]

List of tag keys to remove from the IoT Device Advisor resource.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

ListSuiteRuns

listSuiteRuns_suiteDefinitionId :: Lens' ListSuiteRuns (Maybe Text)

Lists the test suite runs of the specified test suite based on suite definition Id.

listSuiteRuns_suiteDefinitionVersion :: Lens' ListSuiteRuns (Maybe Text)

Must be passed along with suiteDefinitionId. Lists the test suite runs of the specified test suite based on suite definition version.

listSuiteRuns_nextToken :: Lens' ListSuiteRuns (Maybe Text)

A token to retrieve the next set of results.

listSuiteRuns_maxResults :: Lens' ListSuiteRuns (Maybe Natural)

The maximum number of results to return at once.

listSuiteRunsResponse_nextToken :: Lens' ListSuiteRunsResponse (Maybe Text)

A token to retrieve the next set of results.

listSuiteRunsResponse_suiteRunsList :: Lens' ListSuiteRunsResponse (Maybe [SuiteRunInformation])

An array of objects that provide summaries of information about the suite runs in the list.

listSuiteRunsResponse_httpStatus :: Lens' ListSuiteRunsResponse Int

The response's http status code.

GetSuiteRun

getSuiteRun_suiteDefinitionId :: Lens' GetSuiteRun Text

Suite definition Id for the test suite run.

getSuiteRun_suiteRunId :: Lens' GetSuiteRun Text

Suite run Id for the test suite run.

getSuiteRunResponse_status :: Lens' GetSuiteRunResponse (Maybe SuiteRunStatus)

Status for the test suite run.

getSuiteRunResponse_suiteRunConfiguration :: Lens' GetSuiteRunResponse (Maybe SuiteRunConfiguration)

Suite run configuration for the test suite run.

getSuiteRunResponse_startTime :: Lens' GetSuiteRunResponse (Maybe UTCTime)

Date (in Unix epoch time) when the test suite run was started.

getSuiteRunResponse_suiteDefinitionId :: Lens' GetSuiteRunResponse (Maybe Text)

Suite definition Id for the test suite run.

getSuiteRunResponse_suiteDefinitionVersion :: Lens' GetSuiteRunResponse (Maybe Text)

Suite definition version for the test suite run.

getSuiteRunResponse_errorReason :: Lens' GetSuiteRunResponse (Maybe Text)

Error reason for any test suite run failure.

getSuiteRunResponse_testResult :: Lens' GetSuiteRunResponse (Maybe TestResult)

Test results for the test suite run.

getSuiteRunResponse_endTime :: Lens' GetSuiteRunResponse (Maybe UTCTime)

Date (in Unix epoch time) when the test suite run ended.

getSuiteRunResponse_suiteRunArn :: Lens' GetSuiteRunResponse (Maybe Text)

The ARN of the suite run.

getSuiteRunResponse_suiteRunId :: Lens' GetSuiteRunResponse (Maybe Text)

Suite run Id for the test suite run.

getSuiteRunResponse_tags :: Lens' GetSuiteRunResponse (Maybe (HashMap Text Text))

The tags attached to the suite run.

getSuiteRunResponse_httpStatus :: Lens' GetSuiteRunResponse Int

The response's http status code.

Types

DeviceUnderTest

deviceUnderTest_certificateArn :: Lens' DeviceUnderTest (Maybe Text)

Lists devices certificate arn

deviceUnderTest_thingArn :: Lens' DeviceUnderTest (Maybe Text)

Lists devices thing arn

GroupResult

groupResult_tests :: Lens' GroupResult (Maybe [TestCaseRun])

Tests under Group Result.

groupResult_groupId :: Lens' GroupResult (Maybe Text)

Group result Id.

groupResult_groupName :: Lens' GroupResult (Maybe Text)

Group Result Name.

SuiteDefinitionConfiguration

suiteDefinitionConfiguration_suiteDefinitionName :: Lens' SuiteDefinitionConfiguration (Maybe Text)

Gets Suite Definition Configuration name.

suiteDefinitionConfiguration_intendedForQualification :: Lens' SuiteDefinitionConfiguration (Maybe Bool)

Gets the tests intended for qualification in a suite.

suiteDefinitionConfiguration_devicePermissionRoleArn :: Lens' SuiteDefinitionConfiguration (Maybe Text)

Gets device permission arn.

suiteDefinitionConfiguration_devices :: Lens' SuiteDefinitionConfiguration (Maybe [DeviceUnderTest])

Gets the devices configured.

suiteDefinitionConfiguration_rootGroup :: Lens' SuiteDefinitionConfiguration (Maybe Text)

Gets test suite root group.

SuiteDefinitionInformation

suiteDefinitionInformation_createdAt :: Lens' SuiteDefinitionInformation (Maybe UTCTime)

Date (in Unix epoch time) when the test suite was created.

suiteDefinitionInformation_defaultDevices :: Lens' SuiteDefinitionInformation (Maybe [DeviceUnderTest])

Specifies the devices under test for the test suite.

suiteDefinitionInformation_suiteDefinitionId :: Lens' SuiteDefinitionInformation (Maybe Text)

Suite definition Id of the test suite.

suiteDefinitionInformation_suiteDefinitionName :: Lens' SuiteDefinitionInformation (Maybe Text)

Suite name of the test suite.

suiteDefinitionInformation_intendedForQualification :: Lens' SuiteDefinitionInformation (Maybe Bool)

Specifies if the test suite is intended for qualification.

SuiteRunConfiguration

suiteRunConfiguration_primaryDevice :: Lens' SuiteRunConfiguration (Maybe DeviceUnderTest)

Gets the primary device for suite run.

suiteRunConfiguration_selectedTestList :: Lens' SuiteRunConfiguration (Maybe [Text])

Gets test case list.

SuiteRunInformation

suiteRunInformation_status :: Lens' SuiteRunInformation (Maybe SuiteRunStatus)

Status of the suite run.

suiteRunInformation_createdAt :: Lens' SuiteRunInformation (Maybe UTCTime)

Date (in Unix epoch time) when the suite run was created.

suiteRunInformation_passed :: Lens' SuiteRunInformation (Maybe Natural)

Number of test cases that passed in the suite run.

suiteRunInformation_suiteDefinitionId :: Lens' SuiteRunInformation (Maybe Text)

Suite definition Id of the suite run.

suiteRunInformation_suiteDefinitionVersion :: Lens' SuiteRunInformation (Maybe Text)

Suite definition version of the suite run.

suiteRunInformation_startedAt :: Lens' SuiteRunInformation (Maybe UTCTime)

Date (in Unix epoch time) when the suite run was started.

suiteRunInformation_suiteDefinitionName :: Lens' SuiteRunInformation (Maybe Text)

Suite definition name of the suite run.

suiteRunInformation_endAt :: Lens' SuiteRunInformation (Maybe UTCTime)

Date (in Unix epoch time) when the suite run ended.

suiteRunInformation_suiteRunId :: Lens' SuiteRunInformation (Maybe Text)

Suite run Id of the suite run.

suiteRunInformation_failed :: Lens' SuiteRunInformation (Maybe Natural)

Number of test cases that failed in the suite run.

TestCaseRun

testCaseRun_status :: Lens' TestCaseRun (Maybe Status)

Provides test case run status.

testCaseRun_logUrl :: Lens' TestCaseRun (Maybe Text)

Provides test case run log Url.

testCaseRun_startTime :: Lens' TestCaseRun (Maybe UTCTime)

Provides test case run start time.

testCaseRun_testCaseRunId :: Lens' TestCaseRun (Maybe Text)

Provides test case run Id.

testCaseRun_warnings :: Lens' TestCaseRun (Maybe Text)

Provides test case run warnings.

testCaseRun_endTime :: Lens' TestCaseRun (Maybe UTCTime)

Provides test case run end time.

testCaseRun_testCaseDefinitionId :: Lens' TestCaseRun (Maybe Text)

Provides test case run definition Id.

testCaseRun_failure :: Lens' TestCaseRun (Maybe Text)

Provides test case run failure result.

testCaseRun_testCaseDefinitionName :: Lens' TestCaseRun (Maybe Text)

Provides test case run definition Name.

TestResult

testResult_groups :: Lens' TestResult (Maybe [GroupResult])

Show each group of test results.