Network.AWS.IoTDeviceAdvisor.Types

Description

Service Configuration

defaultService :: Service

API version 2020-09-18 of the Amazon IoT Core Device Advisor SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

Sends invalid request exception.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Sends Conflict Exception.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

Sends Internal Failure Exception.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Sends Resource Not Found Exception.

Status

newtype Status

Constructors

Status'

Fields

fromStatus :: Text

Bundled Patterns

pattern Status_CANCELED :: Status

pattern Status_ERROR :: Status

pattern Status_FAIL :: Status

pattern Status_PASS :: Status

pattern Status_PASS_WITH_WARNINGS :: Status

pattern Status_PENDING :: Status

pattern Status_RUNNING :: Status

pattern Status_STOPPED :: Status

pattern Status_STOPPING :: Status

SuiteRunStatus

newtype SuiteRunStatus

Constructors

SuiteRunStatus'

Fields

fromSuiteRunStatus :: Text

Bundled Patterns

pattern SuiteRunStatus_CANCELED :: SuiteRunStatus

pattern SuiteRunStatus_ERROR :: SuiteRunStatus

pattern SuiteRunStatus_FAIL :: SuiteRunStatus

pattern SuiteRunStatus_PASS :: SuiteRunStatus

pattern SuiteRunStatus_PASS_WITH_WARNINGS :: SuiteRunStatus

pattern SuiteRunStatus_PENDING :: SuiteRunStatus

pattern SuiteRunStatus_RUNNING :: SuiteRunStatus

pattern SuiteRunStatus_STOPPED :: SuiteRunStatus

pattern SuiteRunStatus_STOPPING :: SuiteRunStatus

DeviceUnderTest

data DeviceUnderTest

Lists all the devices under test

See: newDeviceUnderTest smart constructor.

Constructors

DeviceUnderTest'

Fields

certificateArn :: Maybe Text

Lists devices certificate arn

thingArn :: Maybe Text

Lists devices thing arn

newDeviceUnderTest :: DeviceUnderTest

Create a value of DeviceUnderTest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateArn:DeviceUnderTest', deviceUnderTest_certificateArn - Lists devices certificate arn

$sel:thingArn:DeviceUnderTest', deviceUnderTest_thingArn - Lists devices thing arn

deviceUnderTest_certificateArn :: Lens' DeviceUnderTest (Maybe Text)

Lists devices certificate arn

deviceUnderTest_thingArn :: Lens' DeviceUnderTest (Maybe Text)

Lists devices thing arn

GroupResult

data GroupResult

Show Group Result.

See: newGroupResult smart constructor.

Constructors

GroupResult'

Fields

tests :: Maybe [TestCaseRun]

Tests under Group Result.

groupId :: Maybe Text

Group result Id.

groupName :: Maybe Text

Group Result Name.

newGroupResult :: GroupResult

Create a value of GroupResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tests:GroupResult', groupResult_tests - Tests under Group Result.

$sel:groupId:GroupResult', groupResult_groupId - Group result Id.

$sel:groupName:GroupResult', groupResult_groupName - Group Result Name.

groupResult_tests :: Lens' GroupResult (Maybe [TestCaseRun])

Tests under Group Result.

groupResult_groupId :: Lens' GroupResult (Maybe Text)

Group result Id.

groupResult_groupName :: Lens' GroupResult (Maybe Text)

Group Result Name.

SuiteDefinitionConfiguration

data SuiteDefinitionConfiguration

Gets Suite Definition Configuration.

See: newSuiteDefinitionConfiguration smart constructor.

Constructors

SuiteDefinitionConfiguration'

Fields

suiteDefinitionName :: Maybe Text

Gets Suite Definition Configuration name.

intendedForQualification :: Maybe Bool

Gets the tests intended for qualification in a suite.

devicePermissionRoleArn :: Maybe Text

Gets device permission arn.

devices :: Maybe [DeviceUnderTest]

Gets the devices configured.

rootGroup :: Maybe Text

Gets test suite root group.

newSuiteDefinitionConfiguration :: SuiteDefinitionConfiguration

Create a value of SuiteDefinitionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suiteDefinitionName:SuiteDefinitionConfiguration', suiteDefinitionConfiguration_suiteDefinitionName - Gets Suite Definition Configuration name.

$sel:intendedForQualification:SuiteDefinitionConfiguration', suiteDefinitionConfiguration_intendedForQualification - Gets the tests intended for qualification in a suite.

$sel:devicePermissionRoleArn:SuiteDefinitionConfiguration', suiteDefinitionConfiguration_devicePermissionRoleArn - Gets device permission arn.

$sel:devices:SuiteDefinitionConfiguration', suiteDefinitionConfiguration_devices - Gets the devices configured.

$sel:rootGroup:SuiteDefinitionConfiguration', suiteDefinitionConfiguration_rootGroup - Gets test suite root group.

suiteDefinitionConfiguration_suiteDefinitionName :: Lens' SuiteDefinitionConfiguration (Maybe Text)

Gets Suite Definition Configuration name.

suiteDefinitionConfiguration_intendedForQualification :: Lens' SuiteDefinitionConfiguration (Maybe Bool)

Gets the tests intended for qualification in a suite.

suiteDefinitionConfiguration_devicePermissionRoleArn :: Lens' SuiteDefinitionConfiguration (Maybe Text)

Gets device permission arn.

suiteDefinitionConfiguration_devices :: Lens' SuiteDefinitionConfiguration (Maybe [DeviceUnderTest])

Gets the devices configured.

suiteDefinitionConfiguration_rootGroup :: Lens' SuiteDefinitionConfiguration (Maybe Text)

Gets test suite root group.

SuiteDefinitionInformation

data SuiteDefinitionInformation

Information about the suite definition.

See: newSuiteDefinitionInformation smart constructor.

Constructors

SuiteDefinitionInformation'

Fields

createdAt :: Maybe POSIX

Date (in Unix epoch time) when the test suite was created.

defaultDevices :: Maybe [DeviceUnderTest]

Specifies the devices under test for the test suite.

suiteDefinitionId :: Maybe Text

Suite definition Id of the test suite.

suiteDefinitionName :: Maybe Text

Suite name of the test suite.

intendedForQualification :: Maybe Bool

Specifies if the test suite is intended for qualification.

newSuiteDefinitionInformation :: SuiteDefinitionInformation

Create a value of SuiteDefinitionInformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAt:SuiteDefinitionInformation', suiteDefinitionInformation_createdAt - Date (in Unix epoch time) when the test suite was created.

$sel:defaultDevices:SuiteDefinitionInformation', suiteDefinitionInformation_defaultDevices - Specifies the devices under test for the test suite.

$sel:suiteDefinitionId:SuiteDefinitionInformation', suiteDefinitionInformation_suiteDefinitionId - Suite definition Id of the test suite.

$sel:suiteDefinitionName:SuiteDefinitionInformation', suiteDefinitionInformation_suiteDefinitionName - Suite name of the test suite.

$sel:intendedForQualification:SuiteDefinitionInformation', suiteDefinitionInformation_intendedForQualification - Specifies if the test suite is intended for qualification.

suiteDefinitionInformation_createdAt :: Lens' SuiteDefinitionInformation (Maybe UTCTime)

Date (in Unix epoch time) when the test suite was created.

suiteDefinitionInformation_defaultDevices :: Lens' SuiteDefinitionInformation (Maybe [DeviceUnderTest])

Specifies the devices under test for the test suite.

suiteDefinitionInformation_suiteDefinitionId :: Lens' SuiteDefinitionInformation (Maybe Text)

Suite definition Id of the test suite.

suiteDefinitionInformation_suiteDefinitionName :: Lens' SuiteDefinitionInformation (Maybe Text)

Suite name of the test suite.

suiteDefinitionInformation_intendedForQualification :: Lens' SuiteDefinitionInformation (Maybe Bool)

Specifies if the test suite is intended for qualification.

SuiteRunConfiguration

data SuiteRunConfiguration

Gets suite run configuration.

See: newSuiteRunConfiguration smart constructor.

Constructors

SuiteRunConfiguration'

Fields

primaryDevice :: Maybe DeviceUnderTest

Gets the primary device for suite run.

selectedTestList :: Maybe [Text]

Gets test case list.

newSuiteRunConfiguration :: SuiteRunConfiguration

Create a value of SuiteRunConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:primaryDevice:SuiteRunConfiguration', suiteRunConfiguration_primaryDevice - Gets the primary device for suite run.

$sel:selectedTestList:SuiteRunConfiguration', suiteRunConfiguration_selectedTestList - Gets test case list.

suiteRunConfiguration_primaryDevice :: Lens' SuiteRunConfiguration (Maybe DeviceUnderTest)

Gets the primary device for suite run.

suiteRunConfiguration_selectedTestList :: Lens' SuiteRunConfiguration (Maybe [Text])

Gets test case list.

SuiteRunInformation

data SuiteRunInformation

Information about the suite run.

See: newSuiteRunInformation smart constructor.

Constructors

SuiteRunInformation'

Fields

status :: Maybe SuiteRunStatus

Status of the suite run.

createdAt :: Maybe POSIX

Date (in Unix epoch time) when the suite run was created.

passed :: Maybe Natural

Number of test cases that passed in the suite run.

suiteDefinitionId :: Maybe Text

Suite definition Id of the suite run.

suiteDefinitionVersion :: Maybe Text

Suite definition version of the suite run.

startedAt :: Maybe POSIX

Date (in Unix epoch time) when the suite run was started.

suiteDefinitionName :: Maybe Text

Suite definition name of the suite run.

endAt :: Maybe POSIX

Date (in Unix epoch time) when the suite run ended.

suiteRunId :: Maybe Text

Suite run Id of the suite run.

failed :: Maybe Natural

Number of test cases that failed in the suite run.

newSuiteRunInformation :: SuiteRunInformation

Create a value of SuiteRunInformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SuiteRunInformation', suiteRunInformation_status - Status of the suite run.

$sel:createdAt:SuiteRunInformation', suiteRunInformation_createdAt - Date (in Unix epoch time) when the suite run was created.

$sel:passed:SuiteRunInformation', suiteRunInformation_passed - Number of test cases that passed in the suite run.

$sel:suiteDefinitionId:SuiteRunInformation', suiteRunInformation_suiteDefinitionId - Suite definition Id of the suite run.

$sel:suiteDefinitionVersion:SuiteRunInformation', suiteRunInformation_suiteDefinitionVersion - Suite definition version of the suite run.

$sel:startedAt:SuiteRunInformation', suiteRunInformation_startedAt - Date (in Unix epoch time) when the suite run was started.

$sel:suiteDefinitionName:SuiteRunInformation', suiteRunInformation_suiteDefinitionName - Suite definition name of the suite run.

$sel:endAt:SuiteRunInformation', suiteRunInformation_endAt - Date (in Unix epoch time) when the suite run ended.

$sel:suiteRunId:SuiteRunInformation', suiteRunInformation_suiteRunId - Suite run Id of the suite run.

$sel:failed:SuiteRunInformation', suiteRunInformation_failed - Number of test cases that failed in the suite run.

suiteRunInformation_status :: Lens' SuiteRunInformation (Maybe SuiteRunStatus)

Status of the suite run.

suiteRunInformation_createdAt :: Lens' SuiteRunInformation (Maybe UTCTime)

Date (in Unix epoch time) when the suite run was created.

suiteRunInformation_passed :: Lens' SuiteRunInformation (Maybe Natural)

Number of test cases that passed in the suite run.

suiteRunInformation_suiteDefinitionId :: Lens' SuiteRunInformation (Maybe Text)

Suite definition Id of the suite run.

suiteRunInformation_suiteDefinitionVersion :: Lens' SuiteRunInformation (Maybe Text)

Suite definition version of the suite run.

suiteRunInformation_startedAt :: Lens' SuiteRunInformation (Maybe UTCTime)

Date (in Unix epoch time) when the suite run was started.

suiteRunInformation_suiteDefinitionName :: Lens' SuiteRunInformation (Maybe Text)

Suite definition name of the suite run.

suiteRunInformation_endAt :: Lens' SuiteRunInformation (Maybe UTCTime)

Date (in Unix epoch time) when the suite run ended.

suiteRunInformation_suiteRunId :: Lens' SuiteRunInformation (Maybe Text)

Suite run Id of the suite run.

suiteRunInformation_failed :: Lens' SuiteRunInformation (Maybe Natural)

Number of test cases that failed in the suite run.

TestCaseRun

data TestCaseRun

Provides test case run.

See: newTestCaseRun smart constructor.

Constructors

TestCaseRun'

Fields

status :: Maybe Status

Provides test case run status.

logUrl :: Maybe Text

Provides test case run log Url.

startTime :: Maybe POSIX

Provides test case run start time.

testCaseRunId :: Maybe Text

Provides test case run Id.

warnings :: Maybe Text

Provides test case run warnings.

endTime :: Maybe POSIX

Provides test case run end time.

testCaseDefinitionId :: Maybe Text

Provides test case run definition Id.

failure :: Maybe Text

Provides test case run failure result.

testCaseDefinitionName :: Maybe Text

Provides test case run definition Name.

newTestCaseRun :: TestCaseRun

Create a value of TestCaseRun with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:TestCaseRun', testCaseRun_status - Provides test case run status.

$sel:logUrl:TestCaseRun', testCaseRun_logUrl - Provides test case run log Url.

$sel:startTime:TestCaseRun', testCaseRun_startTime - Provides test case run start time.

$sel:testCaseRunId:TestCaseRun', testCaseRun_testCaseRunId - Provides test case run Id.

$sel:warnings:TestCaseRun', testCaseRun_warnings - Provides test case run warnings.

$sel:endTime:TestCaseRun', testCaseRun_endTime - Provides test case run end time.

$sel:testCaseDefinitionId:TestCaseRun', testCaseRun_testCaseDefinitionId - Provides test case run definition Id.

$sel:failure:TestCaseRun', testCaseRun_failure - Provides test case run failure result.

$sel:testCaseDefinitionName:TestCaseRun', testCaseRun_testCaseDefinitionName - Provides test case run definition Name.

testCaseRun_status :: Lens' TestCaseRun (Maybe Status)

Provides test case run status.

testCaseRun_logUrl :: Lens' TestCaseRun (Maybe Text)

Provides test case run log Url.

testCaseRun_startTime :: Lens' TestCaseRun (Maybe UTCTime)

Provides test case run start time.

testCaseRun_testCaseRunId :: Lens' TestCaseRun (Maybe Text)

Provides test case run Id.

testCaseRun_warnings :: Lens' TestCaseRun (Maybe Text)

Provides test case run warnings.

testCaseRun_endTime :: Lens' TestCaseRun (Maybe UTCTime)

Provides test case run end time.

testCaseRun_testCaseDefinitionId :: Lens' TestCaseRun (Maybe Text)

Provides test case run definition Id.

testCaseRun_failure :: Lens' TestCaseRun (Maybe Text)

Provides test case run failure result.

testCaseRun_testCaseDefinitionName :: Lens' TestCaseRun (Maybe Text)

Provides test case run definition Name.

TestResult

data TestResult

Show each group result.

See: newTestResult smart constructor.

Constructors

TestResult'

Fields

groups :: Maybe [GroupResult]

Show each group of test results.

newTestResult :: TestResult

Create a value of TestResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groups:TestResult', testResult_groups - Show each group of test results.

testResult_groups :: Lens' TestResult (Maybe [GroupResult])

Show each group of test results.