Network.AWS.IoTDeviceAdvisor.Types.SuiteDefinitionInformation

Description

Documentation

data SuiteDefinitionInformation

Information about the suite definition.

See: newSuiteDefinitionInformation smart constructor.

Constructors

SuiteDefinitionInformation'

Fields

createdAt :: Maybe POSIX

Date (in Unix epoch time) when the test suite was created.

defaultDevices :: Maybe [DeviceUnderTest]

Specifies the devices under test for the test suite.

suiteDefinitionId :: Maybe Text

Suite definition Id of the test suite.

suiteDefinitionName :: Maybe Text

Suite name of the test suite.

intendedForQualification :: Maybe Bool

Specifies if the test suite is intended for qualification.

newSuiteDefinitionInformation :: SuiteDefinitionInformation

Create a value of SuiteDefinitionInformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAt:SuiteDefinitionInformation', suiteDefinitionInformation_createdAt - Date (in Unix epoch time) when the test suite was created.

$sel:defaultDevices:SuiteDefinitionInformation', suiteDefinitionInformation_defaultDevices - Specifies the devices under test for the test suite.

$sel:suiteDefinitionId:SuiteDefinitionInformation', suiteDefinitionInformation_suiteDefinitionId - Suite definition Id of the test suite.

$sel:suiteDefinitionName:SuiteDefinitionInformation', suiteDefinitionInformation_suiteDefinitionName - Suite name of the test suite.

$sel:intendedForQualification:SuiteDefinitionInformation', suiteDefinitionInformation_intendedForQualification - Specifies if the test suite is intended for qualification.

suiteDefinitionInformation_createdAt :: Lens' SuiteDefinitionInformation (Maybe UTCTime)

Date (in Unix epoch time) when the test suite was created.

suiteDefinitionInformation_defaultDevices :: Lens' SuiteDefinitionInformation (Maybe [DeviceUnderTest])

Specifies the devices under test for the test suite.

suiteDefinitionInformation_suiteDefinitionId :: Lens' SuiteDefinitionInformation (Maybe Text)

Suite definition Id of the test suite.

suiteDefinitionInformation_suiteDefinitionName :: Lens' SuiteDefinitionInformation (Maybe Text)

Suite name of the test suite.

suiteDefinitionInformation_intendedForQualification :: Lens' SuiteDefinitionInformation (Maybe Bool)

Specifies if the test suite is intended for qualification.