Network.AWS.IoTDeviceAdvisor.Types.SuiteRunInformation

Description

Documentation

data SuiteRunInformation

Information about the suite run.

See: newSuiteRunInformation smart constructor.

Constructors

SuiteRunInformation'

Fields

status :: Maybe SuiteRunStatus

Status of the suite run.

createdAt :: Maybe POSIX

Date (in Unix epoch time) when the suite run was created.

passed :: Maybe Natural

Number of test cases that passed in the suite run.

suiteDefinitionId :: Maybe Text

Suite definition Id of the suite run.

suiteDefinitionVersion :: Maybe Text

Suite definition version of the suite run.

startedAt :: Maybe POSIX

Date (in Unix epoch time) when the suite run was started.

suiteDefinitionName :: Maybe Text

Suite definition name of the suite run.

endAt :: Maybe POSIX

Date (in Unix epoch time) when the suite run ended.

suiteRunId :: Maybe Text

Suite run Id of the suite run.

failed :: Maybe Natural

Number of test cases that failed in the suite run.

newSuiteRunInformation :: SuiteRunInformation

Create a value of SuiteRunInformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SuiteRunInformation', suiteRunInformation_status - Status of the suite run.

$sel:createdAt:SuiteRunInformation', suiteRunInformation_createdAt - Date (in Unix epoch time) when the suite run was created.

$sel:passed:SuiteRunInformation', suiteRunInformation_passed - Number of test cases that passed in the suite run.

$sel:suiteDefinitionId:SuiteRunInformation', suiteRunInformation_suiteDefinitionId - Suite definition Id of the suite run.

$sel:suiteDefinitionVersion:SuiteRunInformation', suiteRunInformation_suiteDefinitionVersion - Suite definition version of the suite run.

$sel:startedAt:SuiteRunInformation', suiteRunInformation_startedAt - Date (in Unix epoch time) when the suite run was started.

$sel:suiteDefinitionName:SuiteRunInformation', suiteRunInformation_suiteDefinitionName - Suite definition name of the suite run.

$sel:endAt:SuiteRunInformation', suiteRunInformation_endAt - Date (in Unix epoch time) when the suite run ended.

$sel:suiteRunId:SuiteRunInformation', suiteRunInformation_suiteRunId - Suite run Id of the suite run.

$sel:failed:SuiteRunInformation', suiteRunInformation_failed - Number of test cases that failed in the suite run.

suiteRunInformation_status :: Lens' SuiteRunInformation (Maybe SuiteRunStatus)

Status of the suite run.

suiteRunInformation_createdAt :: Lens' SuiteRunInformation (Maybe UTCTime)

Date (in Unix epoch time) when the suite run was created.

suiteRunInformation_passed :: Lens' SuiteRunInformation (Maybe Natural)

Number of test cases that passed in the suite run.

suiteRunInformation_suiteDefinitionId :: Lens' SuiteRunInformation (Maybe Text)

Suite definition Id of the suite run.

suiteRunInformation_suiteDefinitionVersion :: Lens' SuiteRunInformation (Maybe Text)

Suite definition version of the suite run.

suiteRunInformation_startedAt :: Lens' SuiteRunInformation (Maybe UTCTime)

Date (in Unix epoch time) when the suite run was started.

suiteRunInformation_suiteDefinitionName :: Lens' SuiteRunInformation (Maybe Text)

Suite definition name of the suite run.

suiteRunInformation_endAt :: Lens' SuiteRunInformation (Maybe UTCTime)

Date (in Unix epoch time) when the suite run ended.

suiteRunInformation_suiteRunId :: Lens' SuiteRunInformation (Maybe Text)

Suite run Id of the suite run.

suiteRunInformation_failed :: Lens' SuiteRunInformation (Maybe Natural)

Number of test cases that failed in the suite run.