Network.AWS.IoTDeviceAdvisor.UpdateSuiteDefinition

Description

Updates a Device Advisor test suite.

Creating a Request

data UpdateSuiteDefinition

See: newUpdateSuiteDefinition smart constructor.

Constructors

UpdateSuiteDefinition'

Fields

suiteDefinitionConfiguration :: Maybe SuiteDefinitionConfiguration

Updates a Device Advisor test suite with suite definition configuration.

suiteDefinitionId :: Text

Suite definition Id of the test suite to be updated.

newUpdateSuiteDefinition

Arguments

:: Text
$sel:suiteDefinitionId:UpdateSuiteDefinition'

-> UpdateSuiteDefinition

Create a value of UpdateSuiteDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:suiteDefinitionConfiguration:UpdateSuiteDefinition', updateSuiteDefinition_suiteDefinitionConfiguration - Updates a Device Advisor test suite with suite definition configuration.

$sel:suiteDefinitionId:UpdateSuiteDefinition', updateSuiteDefinition_suiteDefinitionId - Suite definition Id of the test suite to be updated.

Request Lenses

updateSuiteDefinition_suiteDefinitionConfiguration :: Lens' UpdateSuiteDefinition (Maybe SuiteDefinitionConfiguration)

Updates a Device Advisor test suite with suite definition configuration.

updateSuiteDefinition_suiteDefinitionId :: Lens' UpdateSuiteDefinition Text

Suite definition Id of the test suite to be updated.

Destructuring the Response

data UpdateSuiteDefinitionResponse

See: newUpdateSuiteDefinitionResponse smart constructor.

Constructors

UpdateSuiteDefinitionResponse'

Fields

lastUpdatedAt :: Maybe POSIX

Timestamp of when the test suite was updated.

createdAt :: Maybe POSIX

Timestamp of when the test suite was created.

suiteDefinitionArn :: Maybe Text

Amazon Resource name of the updated test suite.

suiteDefinitionId :: Maybe Text

Suite definition Id of the updated test suite.

suiteDefinitionVersion :: Maybe Text

Suite definition version of the updated test suite.

suiteDefinitionName :: Maybe Text

Suite definition name of the updated test suite.

httpStatus :: Int

The response's http status code.

newUpdateSuiteDefinitionResponse

Arguments

:: Int
$sel:httpStatus:UpdateSuiteDefinitionResponse'

-> UpdateSuiteDefinitionResponse

Create a value of UpdateSuiteDefinitionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedAt:UpdateSuiteDefinitionResponse', updateSuiteDefinitionResponse_lastUpdatedAt - Timestamp of when the test suite was updated.

$sel:createdAt:UpdateSuiteDefinitionResponse', updateSuiteDefinitionResponse_createdAt - Timestamp of when the test suite was created.

$sel:suiteDefinitionArn:UpdateSuiteDefinitionResponse', updateSuiteDefinitionResponse_suiteDefinitionArn - Amazon Resource name of the updated test suite.

$sel:suiteDefinitionId:UpdateSuiteDefinition', updateSuiteDefinitionResponse_suiteDefinitionId - Suite definition Id of the updated test suite.

$sel:suiteDefinitionVersion:UpdateSuiteDefinitionResponse', updateSuiteDefinitionResponse_suiteDefinitionVersion - Suite definition version of the updated test suite.

$sel:suiteDefinitionName:UpdateSuiteDefinitionResponse', updateSuiteDefinitionResponse_suiteDefinitionName - Suite definition name of the updated test suite.

$sel:httpStatus:UpdateSuiteDefinitionResponse', updateSuiteDefinitionResponse_httpStatus - The response's http status code.

Response Lenses

updateSuiteDefinitionResponse_lastUpdatedAt :: Lens' UpdateSuiteDefinitionResponse (Maybe UTCTime)

Timestamp of when the test suite was updated.

updateSuiteDefinitionResponse_createdAt :: Lens' UpdateSuiteDefinitionResponse (Maybe UTCTime)

Timestamp of when the test suite was created.

updateSuiteDefinitionResponse_suiteDefinitionArn :: Lens' UpdateSuiteDefinitionResponse (Maybe Text)

Amazon Resource name of the updated test suite.

updateSuiteDefinitionResponse_suiteDefinitionId :: Lens' UpdateSuiteDefinitionResponse (Maybe Text)

Suite definition Id of the updated test suite.

updateSuiteDefinitionResponse_suiteDefinitionVersion :: Lens' UpdateSuiteDefinitionResponse (Maybe Text)

Suite definition version of the updated test suite.

updateSuiteDefinitionResponse_suiteDefinitionName :: Lens' UpdateSuiteDefinitionResponse (Maybe Text)

Suite definition name of the updated test suite.

updateSuiteDefinitionResponse_httpStatus :: Lens' UpdateSuiteDefinitionResponse Int

The response's http status code.