Network.AWS.IoTEvents.CreateDetectorModel

Description

Creates a detector model.

Creating a Request

data CreateDetectorModel

See: newCreateDetectorModel smart constructor.

Constructors

CreateDetectorModel'

Fields

detectorModelDescription :: Maybe Text

A brief description of the detector model.

key :: Maybe Text

The input attribute key used to identify a device or system to create a detector (an instance of the detector model) and then to route each input received to the appropriate detector (instance). This parameter uses a JSON-path expression in the message payload of each input to specify the attribute-value pair that is used to identify the device associated with the input.

evaluationMethod :: Maybe EvaluationMethod

Information about the order in which events are evaluated and how actions are executed.

tags :: Maybe [Tag]

Metadata that can be used to manage the detector model.

detectorModelName :: Text

The name of the detector model.

detectorModelDefinition :: DetectorModelDefinition

Information that defines how the detectors operate.

roleArn :: Text

The ARN of the role that grants permission to AWS IoT Events to perform its operations.

newCreateDetectorModel

Arguments

:: Text
$sel:detectorModelName:CreateDetectorModel'

-> DetectorModelDefinition
$sel:detectorModelDefinition:CreateDetectorModel'

-> Text
$sel:roleArn:CreateDetectorModel'

-> CreateDetectorModel

Create a value of CreateDetectorModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:detectorModelDescription:CreateDetectorModel', createDetectorModel_detectorModelDescription - A brief description of the detector model.

$sel:key:CreateDetectorModel', createDetectorModel_key - The input attribute key used to identify a device or system to create a detector (an instance of the detector model) and then to route each input received to the appropriate detector (instance). This parameter uses a JSON-path expression in the message payload of each input to specify the attribute-value pair that is used to identify the device associated with the input.

$sel:evaluationMethod:CreateDetectorModel', createDetectorModel_evaluationMethod - Information about the order in which events are evaluated and how actions are executed.

$sel:tags:CreateDetectorModel', createDetectorModel_tags - Metadata that can be used to manage the detector model.

$sel:detectorModelName:CreateDetectorModel', createDetectorModel_detectorModelName - The name of the detector model.

$sel:detectorModelDefinition:CreateDetectorModel', createDetectorModel_detectorModelDefinition - Information that defines how the detectors operate.

$sel:roleArn:CreateDetectorModel', createDetectorModel_roleArn - The ARN of the role that grants permission to AWS IoT Events to perform its operations.

Request Lenses

createDetectorModel_detectorModelDescription :: Lens' CreateDetectorModel (Maybe Text)

A brief description of the detector model.

createDetectorModel_key :: Lens' CreateDetectorModel (Maybe Text)

The input attribute key used to identify a device or system to create a detector (an instance of the detector model) and then to route each input received to the appropriate detector (instance). This parameter uses a JSON-path expression in the message payload of each input to specify the attribute-value pair that is used to identify the device associated with the input.

createDetectorModel_evaluationMethod :: Lens' CreateDetectorModel (Maybe EvaluationMethod)

Information about the order in which events are evaluated and how actions are executed.

createDetectorModel_tags :: Lens' CreateDetectorModel (Maybe [Tag])

Metadata that can be used to manage the detector model.

createDetectorModel_detectorModelName :: Lens' CreateDetectorModel Text

The name of the detector model.

createDetectorModel_detectorModelDefinition :: Lens' CreateDetectorModel DetectorModelDefinition

Information that defines how the detectors operate.

createDetectorModel_roleArn :: Lens' CreateDetectorModel Text

The ARN of the role that grants permission to AWS IoT Events to perform its operations.

Destructuring the Response

data CreateDetectorModelResponse

See: newCreateDetectorModelResponse smart constructor.

Constructors

CreateDetectorModelResponse'

Fields

detectorModelConfiguration :: Maybe DetectorModelConfiguration

Information about how the detector model is configured.

httpStatus :: Int

The response's http status code.

newCreateDetectorModelResponse

Arguments

:: Int
$sel:httpStatus:CreateDetectorModelResponse'

-> CreateDetectorModelResponse

Create a value of CreateDetectorModelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:detectorModelConfiguration:CreateDetectorModelResponse', createDetectorModelResponse_detectorModelConfiguration - Information about how the detector model is configured.

$sel:httpStatus:CreateDetectorModelResponse', createDetectorModelResponse_httpStatus - The response's http status code.

Response Lenses

createDetectorModelResponse_detectorModelConfiguration :: Lens' CreateDetectorModelResponse (Maybe DetectorModelConfiguration)

Information about how the detector model is configured.

createDetectorModelResponse_httpStatus :: Lens' CreateDetectorModelResponse Int

The response's http status code.