Network.AWS.IoTEvents.CreateInput

Description

Creates an input.

Creating a Request

data CreateInput

See: newCreateInput smart constructor.

Constructors

CreateInput'

Fields

inputDescription :: Maybe Text

A brief description of the input.

tags :: Maybe [Tag]

Metadata that can be used to manage the input.

inputName :: Text

The name you want to give to the input.

inputDefinition :: InputDefinition

The definition of the input.

newCreateInput

Arguments

:: Text
$sel:inputName:CreateInput'

-> InputDefinition
$sel:inputDefinition:CreateInput'

-> CreateInput

Create a value of CreateInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputDescription:CreateInput', createInput_inputDescription - A brief description of the input.

$sel:tags:CreateInput', createInput_tags - Metadata that can be used to manage the input.

$sel:inputName:CreateInput', createInput_inputName - The name you want to give to the input.

$sel:inputDefinition:CreateInput', createInput_inputDefinition - The definition of the input.

Request Lenses

createInput_inputDescription :: Lens' CreateInput (Maybe Text)

A brief description of the input.

createInput_tags :: Lens' CreateInput (Maybe [Tag])

Metadata that can be used to manage the input.

createInput_inputName :: Lens' CreateInput Text

The name you want to give to the input.

createInput_inputDefinition :: Lens' CreateInput InputDefinition

The definition of the input.

Destructuring the Response

data CreateInputResponse

See: newCreateInputResponse smart constructor.

Constructors

CreateInputResponse'

Fields

inputConfiguration :: Maybe InputConfiguration

Information about the configuration of the input.

httpStatus :: Int

The response's http status code.

newCreateInputResponse

Arguments

:: Int
$sel:httpStatus:CreateInputResponse'

-> CreateInputResponse

Create a value of CreateInputResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputConfiguration:CreateInputResponse', createInputResponse_inputConfiguration - Information about the configuration of the input.

$sel:httpStatus:CreateInputResponse', createInputResponse_httpStatus - The response's http status code.

Response Lenses

createInputResponse_inputConfiguration :: Lens' CreateInputResponse (Maybe InputConfiguration)

Information about the configuration of the input.

createInputResponse_httpStatus :: Lens' CreateInputResponse Int

The response's http status code.