Network.AWS.IoTEvents.ListDetectorModels

Description

Lists the detector models you have created. Only the metadata associated with each detector model is returned.

Creating a Request

data ListDetectorModels

See: newListDetectorModels smart constructor.

Constructors

ListDetectorModels'

Fields

nextToken :: Maybe Text

The token that you can use to return the next set of results.

maxResults :: Maybe Natural

The maximum number of results to be returned per request.

newListDetectorModels :: ListDetectorModels

Create a value of ListDetectorModels with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDetectorModels', listDetectorModels_nextToken - The token that you can use to return the next set of results.

$sel:maxResults:ListDetectorModels', listDetectorModels_maxResults - The maximum number of results to be returned per request.

Request Lenses

listDetectorModels_nextToken :: Lens' ListDetectorModels (Maybe Text)

The token that you can use to return the next set of results.

listDetectorModels_maxResults :: Lens' ListDetectorModels (Maybe Natural)

The maximum number of results to be returned per request.

Destructuring the Response

data ListDetectorModelsResponse

See: newListDetectorModelsResponse smart constructor.

Constructors

ListDetectorModelsResponse'

Fields

nextToken :: Maybe Text

The token that you can use to return the next set of results, or null if there are no more results.

detectorModelSummaries :: Maybe [DetectorModelSummary]

Summary information about the detector models.

httpStatus :: Int

The response's http status code.

newListDetectorModelsResponse

Arguments

:: Int
$sel:httpStatus:ListDetectorModelsResponse'

-> ListDetectorModelsResponse

Create a value of ListDetectorModelsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDetectorModels', listDetectorModelsResponse_nextToken - The token that you can use to return the next set of results, or null if there are no more results.

$sel:detectorModelSummaries:ListDetectorModelsResponse', listDetectorModelsResponse_detectorModelSummaries - Summary information about the detector models.

$sel:httpStatus:ListDetectorModelsResponse', listDetectorModelsResponse_httpStatus - The response's http status code.

Response Lenses

listDetectorModelsResponse_nextToken :: Lens' ListDetectorModelsResponse (Maybe Text)

The token that you can use to return the next set of results, or null if there are no more results.

listDetectorModelsResponse_detectorModelSummaries :: Lens' ListDetectorModelsResponse (Maybe [DetectorModelSummary])

Summary information about the detector models.

listDetectorModelsResponse_httpStatus :: Lens' ListDetectorModelsResponse Int

The response's http status code.