Network.AWS.IoTEvents.Types.Action

Description

Documentation

data Action

An action to be performed when the condition is TRUE.

See: newAction smart constructor.

Constructors

Action'

Fields

iotTopicPublish :: Maybe IotTopicPublishAction

Publishes an MQTT message with the given topic to the AWS IoT message broker.

dynamoDBv2 :: Maybe DynamoDBv2Action

Writes to the DynamoDB table that you created. The default action payload contains all attribute-value pairs that have the information about the detector model instance and the event that triggered the action. You can customize the payload. A separate column of the DynamoDB table receives one attribute-value pair in the payload that you specify. For more information, see Actions in AWS IoT Events Developer Guide.

resetTimer :: Maybe ResetTimerAction

Information needed to reset the timer.

setTimer :: Maybe SetTimerAction

Information needed to set the timer.

sns :: Maybe SNSTopicPublishAction

Sends an Amazon SNS message.

clearTimer :: Maybe ClearTimerAction

Information needed to clear the timer.

dynamoDB :: Maybe DynamoDBAction

Writes to the DynamoDB table that you created. The default action payload contains all attribute-value pairs that have the information about the detector model instance and the event that triggered the action. You can customize the payload. One column of the DynamoDB table receives all attribute-value pairs in the payload that you specify. For more information, see Actions in AWS IoT Events Developer Guide.

firehose :: Maybe FirehoseAction

Sends information about the detector model instance and the event that triggered the action to an Amazon Kinesis Data Firehose delivery stream.

iotSiteWise :: Maybe IotSiteWiseAction

Sends information about the detector model instance and the event that triggered the action to an asset property in AWS IoT SiteWise .

setVariable :: Maybe SetVariableAction

Sets a variable to a specified value.

lambda :: Maybe LambdaAction

Calls a Lambda function, passing in information about the detector model instance and the event that triggered the action.

iotEvents :: Maybe IotEventsAction

Sends AWS IoT Events input, which passes information about the detector model instance and the event that triggered the action.

sqs :: Maybe SqsAction

Sends information about the detector model instance and the event that triggered the action to an Amazon SQS queue.

newAction :: Action

Create a value of Action with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iotTopicPublish:Action', action_iotTopicPublish - Publishes an MQTT message with the given topic to the AWS IoT message broker.

$sel:dynamoDBv2:Action', action_dynamoDBv2 - Writes to the DynamoDB table that you created. The default action payload contains all attribute-value pairs that have the information about the detector model instance and the event that triggered the action. You can customize the payload. A separate column of the DynamoDB table receives one attribute-value pair in the payload that you specify. For more information, see Actions in AWS IoT Events Developer Guide.

$sel:resetTimer:Action', action_resetTimer - Information needed to reset the timer.

$sel:setTimer:Action', action_setTimer - Information needed to set the timer.

$sel:sns:Action', action_sns - Sends an Amazon SNS message.

$sel:clearTimer:Action', action_clearTimer - Information needed to clear the timer.

$sel:dynamoDB:Action', action_dynamoDB - Writes to the DynamoDB table that you created. The default action payload contains all attribute-value pairs that have the information about the detector model instance and the event that triggered the action. You can customize the payload. One column of the DynamoDB table receives all attribute-value pairs in the payload that you specify. For more information, see Actions in AWS IoT Events Developer Guide.

$sel:firehose:Action', action_firehose - Sends information about the detector model instance and the event that triggered the action to an Amazon Kinesis Data Firehose delivery stream.

$sel:iotSiteWise:Action', action_iotSiteWise - Sends information about the detector model instance and the event that triggered the action to an asset property in AWS IoT SiteWise .

$sel:setVariable:Action', action_setVariable - Sets a variable to a specified value.

$sel:lambda:Action', action_lambda - Calls a Lambda function, passing in information about the detector model instance and the event that triggered the action.

$sel:iotEvents:Action', action_iotEvents - Sends AWS IoT Events input, which passes information about the detector model instance and the event that triggered the action.

$sel:sqs:Action', action_sqs - Sends information about the detector model instance and the event that triggered the action to an Amazon SQS queue.

action_iotTopicPublish :: Lens' Action (Maybe IotTopicPublishAction)

Publishes an MQTT message with the given topic to the AWS IoT message broker.

action_dynamoDBv2 :: Lens' Action (Maybe DynamoDBv2Action)

Writes to the DynamoDB table that you created. The default action payload contains all attribute-value pairs that have the information about the detector model instance and the event that triggered the action. You can customize the payload. A separate column of the DynamoDB table receives one attribute-value pair in the payload that you specify. For more information, see Actions in AWS IoT Events Developer Guide.

action_resetTimer :: Lens' Action (Maybe ResetTimerAction)

Information needed to reset the timer.

action_setTimer :: Lens' Action (Maybe SetTimerAction)

Information needed to set the timer.

action_sns :: Lens' Action (Maybe SNSTopicPublishAction)

Sends an Amazon SNS message.

action_clearTimer :: Lens' Action (Maybe ClearTimerAction)

Information needed to clear the timer.

action_dynamoDB :: Lens' Action (Maybe DynamoDBAction)

Writes to the DynamoDB table that you created. The default action payload contains all attribute-value pairs that have the information about the detector model instance and the event that triggered the action. You can customize the payload. One column of the DynamoDB table receives all attribute-value pairs in the payload that you specify. For more information, see Actions in AWS IoT Events Developer Guide.

action_firehose :: Lens' Action (Maybe FirehoseAction)

Sends information about the detector model instance and the event that triggered the action to an Amazon Kinesis Data Firehose delivery stream.

action_iotSiteWise :: Lens' Action (Maybe IotSiteWiseAction)

Sends information about the detector model instance and the event that triggered the action to an asset property in AWS IoT SiteWise .

action_setVariable :: Lens' Action (Maybe SetVariableAction)

Sets a variable to a specified value.

action_lambda :: Lens' Action (Maybe LambdaAction)

Calls a Lambda function, passing in information about the detector model instance and the event that triggered the action.

action_iotEvents :: Lens' Action (Maybe IotEventsAction)

Sends AWS IoT Events input, which passes information about the detector model instance and the event that triggered the action.

action_sqs :: Lens' Action (Maybe SqsAction)

Sends information about the detector model instance and the event that triggered the action to an Amazon SQS queue.