Network.AWS.IoTEvents.Types.AlarmModelSummary

Description

Documentation

data AlarmModelSummary

Contains a summary of an alarm model.

See: newAlarmModelSummary smart constructor.

Constructors

AlarmModelSummary'

Fields

creationTime :: Maybe POSIX

The time the alarm model was created, in the Unix epoch format.

alarmModelName :: Maybe Text

The name of the alarm model.

alarmModelDescription :: Maybe Text

The description of the alarm model.

newAlarmModelSummary :: AlarmModelSummary

Create a value of AlarmModelSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:AlarmModelSummary', alarmModelSummary_creationTime - The time the alarm model was created, in the Unix epoch format.

$sel:alarmModelName:AlarmModelSummary', alarmModelSummary_alarmModelName - The name of the alarm model.

$sel:alarmModelDescription:AlarmModelSummary', alarmModelSummary_alarmModelDescription - The description of the alarm model.

alarmModelSummary_creationTime :: Lens' AlarmModelSummary (Maybe UTCTime)

The time the alarm model was created, in the Unix epoch format.

alarmModelSummary_alarmModelName :: Lens' AlarmModelSummary (Maybe Text)

The name of the alarm model.

alarmModelSummary_alarmModelDescription :: Lens' AlarmModelSummary (Maybe Text)

The description of the alarm model.