Network.AWS.IoTEvents.Types.AssetPropertyValue

Description

Documentation

data AssetPropertyValue

A structure that contains value information. For more information, see AssetPropertyValue in the AWS IoT SiteWise API Reference.

You must use expressions for all parameters in AssetPropertyValue. The expressions accept literals, operators, functions, references, and substitution templates.

Examples

For literal values, the expressions must contain single quotes. For example, the value for the quality parameter can be 'GOOD'.

For references, you must specify either variables or input values. For example, the value for the quality parameter can be $input.TemperatureInput.sensorData.quality.

For more information, see Expressions in the AWS IoT Events Developer Guide.

See: newAssetPropertyValue smart constructor.

Constructors

AssetPropertyValue'

Fields

value :: Maybe AssetPropertyVariant

The value to send to an asset property.

quality :: Maybe Text

The quality of the asset property value. The value must be 'GOOD', 'BAD', or 'UNCERTAIN'.

timestamp :: Maybe AssetPropertyTimestamp

The timestamp associated with the asset property value. The default is the current event time.

newAssetPropertyValue :: AssetPropertyValue

Create a value of AssetPropertyValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:AssetPropertyValue', assetPropertyValue_value - The value to send to an asset property.

$sel:quality:AssetPropertyValue', assetPropertyValue_quality - The quality of the asset property value. The value must be 'GOOD', 'BAD', or 'UNCERTAIN'.

$sel:timestamp:AssetPropertyValue', assetPropertyValue_timestamp - The timestamp associated with the asset property value. The default is the current event time.

assetPropertyValue_value :: Lens' AssetPropertyValue (Maybe AssetPropertyVariant)

The value to send to an asset property.

assetPropertyValue_quality :: Lens' AssetPropertyValue (Maybe Text)

The quality of the asset property value. The value must be 'GOOD', 'BAD', or 'UNCERTAIN'.

assetPropertyValue_timestamp :: Lens' AssetPropertyValue (Maybe AssetPropertyTimestamp)

The timestamp associated with the asset property value. The default is the current event time.