Network.AWS.IoTEvents.Types.DetectorModelConfiguration

Description

Documentation

data DetectorModelConfiguration

Information about how the detector model is configured.

See: newDetectorModelConfiguration smart constructor.

Constructors

DetectorModelConfiguration'

Fields

creationTime :: Maybe POSIX

The time the detector model was created.

status :: Maybe DetectorModelVersionStatus

The status of the detector model.

detectorModelName :: Maybe Text

The name of the detector model.

detectorModelArn :: Maybe Text

The ARN of the detector model.

detectorModelDescription :: Maybe Text

A brief description of the detector model.

key :: Maybe Text

The value used to identify a detector instance. When a device or system sends input, a new detector instance with a unique key value is created. AWS IoT Events can continue to route input to its corresponding detector instance based on this identifying information.

This parameter uses a JSON-path expression to select the attribute-value pair in the message payload that is used for identification. To route the message to the correct detector instance, the device must send a message payload that contains the same attribute-value.

detectorModelVersion :: Maybe Text

The version of the detector model.

lastUpdateTime :: Maybe POSIX

The time the detector model was last updated.

evaluationMethod :: Maybe EvaluationMethod

Information about the order in which events are evaluated and how actions are executed.

roleArn :: Maybe Text

The ARN of the role that grants permission to AWS IoT Events to perform its operations.

newDetectorModelConfiguration :: DetectorModelConfiguration

Create a value of DetectorModelConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DetectorModelConfiguration', detectorModelConfiguration_creationTime - The time the detector model was created.

$sel:status:DetectorModelConfiguration', detectorModelConfiguration_status - The status of the detector model.

$sel:detectorModelName:DetectorModelConfiguration', detectorModelConfiguration_detectorModelName - The name of the detector model.

$sel:detectorModelArn:DetectorModelConfiguration', detectorModelConfiguration_detectorModelArn - The ARN of the detector model.

$sel:detectorModelDescription:DetectorModelConfiguration', detectorModelConfiguration_detectorModelDescription - A brief description of the detector model.

$sel:key:DetectorModelConfiguration', detectorModelConfiguration_key - The value used to identify a detector instance. When a device or system sends input, a new detector instance with a unique key value is created. AWS IoT Events can continue to route input to its corresponding detector instance based on this identifying information.

This parameter uses a JSON-path expression to select the attribute-value pair in the message payload that is used for identification. To route the message to the correct detector instance, the device must send a message payload that contains the same attribute-value.

$sel:detectorModelVersion:DetectorModelConfiguration', detectorModelConfiguration_detectorModelVersion - The version of the detector model.

$sel:lastUpdateTime:DetectorModelConfiguration', detectorModelConfiguration_lastUpdateTime - The time the detector model was last updated.

$sel:evaluationMethod:DetectorModelConfiguration', detectorModelConfiguration_evaluationMethod - Information about the order in which events are evaluated and how actions are executed.

$sel:roleArn:DetectorModelConfiguration', detectorModelConfiguration_roleArn - The ARN of the role that grants permission to AWS IoT Events to perform its operations.

detectorModelConfiguration_creationTime :: Lens' DetectorModelConfiguration (Maybe UTCTime)

The time the detector model was created.

detectorModelConfiguration_status :: Lens' DetectorModelConfiguration (Maybe DetectorModelVersionStatus)

The status of the detector model.

detectorModelConfiguration_detectorModelName :: Lens' DetectorModelConfiguration (Maybe Text)

The name of the detector model.

detectorModelConfiguration_detectorModelArn :: Lens' DetectorModelConfiguration (Maybe Text)

The ARN of the detector model.

detectorModelConfiguration_detectorModelDescription :: Lens' DetectorModelConfiguration (Maybe Text)

A brief description of the detector model.

detectorModelConfiguration_key :: Lens' DetectorModelConfiguration (Maybe Text)

The value used to identify a detector instance. When a device or system sends input, a new detector instance with a unique key value is created. AWS IoT Events can continue to route input to its corresponding detector instance based on this identifying information.

This parameter uses a JSON-path expression to select the attribute-value pair in the message payload that is used for identification. To route the message to the correct detector instance, the device must send a message payload that contains the same attribute-value.

detectorModelConfiguration_detectorModelVersion :: Lens' DetectorModelConfiguration (Maybe Text)

The version of the detector model.

detectorModelConfiguration_lastUpdateTime :: Lens' DetectorModelConfiguration (Maybe UTCTime)

The time the detector model was last updated.

detectorModelConfiguration_evaluationMethod :: Lens' DetectorModelConfiguration (Maybe EvaluationMethod)

Information about the order in which events are evaluated and how actions are executed.

detectorModelConfiguration_roleArn :: Lens' DetectorModelConfiguration (Maybe Text)

The ARN of the role that grants permission to AWS IoT Events to perform its operations.