Network.AWS.IoTEvents.Types.DetectorModelVersionSummary

Description

Documentation

data DetectorModelVersionSummary

Information about the detector model version.

See: newDetectorModelVersionSummary smart constructor.

Constructors

DetectorModelVersionSummary'

Fields

creationTime :: Maybe POSIX

The time the detector model version was created.

status :: Maybe DetectorModelVersionStatus

The status of the detector model version.

detectorModelName :: Maybe Text

The name of the detector model.

detectorModelArn :: Maybe Text

The ARN of the detector model version.

detectorModelVersion :: Maybe Text

The ID of the detector model version.

lastUpdateTime :: Maybe POSIX

The last time the detector model version was updated.

evaluationMethod :: Maybe EvaluationMethod

Information about the order in which events are evaluated and how actions are executed.

roleArn :: Maybe Text

The ARN of the role that grants the detector model permission to perform its tasks.

newDetectorModelVersionSummary :: DetectorModelVersionSummary

Create a value of DetectorModelVersionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DetectorModelVersionSummary', detectorModelVersionSummary_creationTime - The time the detector model version was created.

$sel:status:DetectorModelVersionSummary', detectorModelVersionSummary_status - The status of the detector model version.

$sel:detectorModelName:DetectorModelVersionSummary', detectorModelVersionSummary_detectorModelName - The name of the detector model.

$sel:detectorModelArn:DetectorModelVersionSummary', detectorModelVersionSummary_detectorModelArn - The ARN of the detector model version.

$sel:detectorModelVersion:DetectorModelVersionSummary', detectorModelVersionSummary_detectorModelVersion - The ID of the detector model version.

$sel:lastUpdateTime:DetectorModelVersionSummary', detectorModelVersionSummary_lastUpdateTime - The last time the detector model version was updated.

$sel:evaluationMethod:DetectorModelVersionSummary', detectorModelVersionSummary_evaluationMethod - Information about the order in which events are evaluated and how actions are executed.

$sel:roleArn:DetectorModelVersionSummary', detectorModelVersionSummary_roleArn - The ARN of the role that grants the detector model permission to perform its tasks.

detectorModelVersionSummary_creationTime :: Lens' DetectorModelVersionSummary (Maybe UTCTime)

The time the detector model version was created.

detectorModelVersionSummary_status :: Lens' DetectorModelVersionSummary (Maybe DetectorModelVersionStatus)

The status of the detector model version.

detectorModelVersionSummary_detectorModelName :: Lens' DetectorModelVersionSummary (Maybe Text)

The name of the detector model.

detectorModelVersionSummary_detectorModelArn :: Lens' DetectorModelVersionSummary (Maybe Text)

The ARN of the detector model version.

detectorModelVersionSummary_detectorModelVersion :: Lens' DetectorModelVersionSummary (Maybe Text)

The ID of the detector model version.

detectorModelVersionSummary_lastUpdateTime :: Lens' DetectorModelVersionSummary (Maybe UTCTime)

The last time the detector model version was updated.

detectorModelVersionSummary_evaluationMethod :: Lens' DetectorModelVersionSummary (Maybe EvaluationMethod)

Information about the order in which events are evaluated and how actions are executed.

detectorModelVersionSummary_roleArn :: Lens' DetectorModelVersionSummary (Maybe Text)

The ARN of the role that grants the detector model permission to perform its tasks.