Network.AWS.IoTEvents.Types.DynamoDBv2Action

Description

Documentation

data DynamoDBv2Action

Defines an action to write to the Amazon DynamoDB table that you created. The default action payload contains all the information about the detector model instance and the event that triggered the action. You can customize the payload. A separate column of the DynamoDB table receives one attribute-value pair in the payload that you specify.

You must use expressions for all parameters in DynamoDBv2Action. The expressions accept literals, operators, functions, references, and substitution templates.

Examples

For literal values, the expressions must contain single quotes. For example, the value for the tableName parameter can be 'GreenhouseTemperatureTable'.

For references, you must specify either variables or input values. For example, the value for the tableName parameter can be $variable.ddbtableName.

For a substitution template, you must use ${}, and the template must be in single quotes. A substitution template can also contain a combination of literals, operators, functions, references, and substitution templates.

In the following example, the value for the contentExpression parameter in Payload uses a substitution template.

'{\"sensorID\": \"${$input.GreenhouseInput.sensor_id}\", \"temperature\": \"${$input.GreenhouseInput.temperature * 9 / 5 + 32}\"}'

For a string concatenation, you must use +. A string concatenation can also contain a combination of literals, operators, functions, references, and substitution templates.

In the following example, the value for the tableName parameter uses a string concatenation.

'GreenhouseTemperatureTable ' + $input.GreenhouseInput.date

For more information, see Expressions in the AWS IoT Events Developer Guide.

The value for the type parameter in Payload must be JSON.

See: newDynamoDBv2Action smart constructor.

Constructors

DynamoDBv2Action'

Fields

payload :: Maybe Payload

tableName :: Text

The name of the DynamoDB table.

newDynamoDBv2Action

Arguments

:: Text
$sel:tableName:DynamoDBv2Action'

-> DynamoDBv2Action

Create a value of DynamoDBv2Action with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:payload:DynamoDBv2Action', dynamoDBv2Action_payload - Undocumented member.

$sel:tableName:DynamoDBv2Action', dynamoDBv2Action_tableName - The name of the DynamoDB table.

dynamoDBv2Action_payload :: Lens' DynamoDBv2Action (Maybe Payload)

Undocumented member.

dynamoDBv2Action_tableName :: Lens' DynamoDBv2Action Text

The name of the DynamoDB table.