Network.AWS.IoTEvents.Types.SMSConfiguration

Description

Documentation

data SMSConfiguration

Contains the configuration information of SMS notifications.

See: newSMSConfiguration smart constructor.

Constructors

SMSConfiguration'

Fields

additionalMessage :: Maybe Text

The message that you want to send. The message can be up to 200 characters.

senderId :: Maybe Text

The sender ID.

recipients :: NonEmpty RecipientDetail

Specifies one or more recipients who receive the message.

You must add the users that receive SMS messages to your AWS SSO store.

newSMSConfiguration

Arguments

:: NonEmpty RecipientDetail
$sel:recipients:SMSConfiguration'

-> SMSConfiguration

Create a value of SMSConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:additionalMessage:SMSConfiguration', sMSConfiguration_additionalMessage - The message that you want to send. The message can be up to 200 characters.

$sel:senderId:SMSConfiguration', sMSConfiguration_senderId - The sender ID.

$sel:recipients:SMSConfiguration', sMSConfiguration_recipients - Specifies one or more recipients who receive the message.

You must add the users that receive SMS messages to your AWS SSO store.

sMSConfiguration_additionalMessage :: Lens' SMSConfiguration (Maybe Text)

The message that you want to send. The message can be up to 200 characters.

sMSConfiguration_senderId :: Lens' SMSConfiguration (Maybe Text)

The sender ID.

sMSConfiguration_recipients :: Lens' SMSConfiguration (NonEmpty RecipientDetail)

Specifies one or more recipients who receive the message.

You must add the users that receive SMS messages to your AWS SSO store.