Network.AWS.IoTEvents.Types.SimpleRule

Description

Documentation

data SimpleRule

A rule that compares an input property value to a threshold value with a comparison operator.

See: newSimpleRule smart constructor.

Constructors

SimpleRule'

Fields

inputProperty :: Text

The value on the left side of the comparison operator. You can specify an AWS IoT Events input attribute as an input property.

comparisonOperator :: ComparisonOperator

The comparison operator.

threshold :: Text

The value on the right side of the comparison operator. You can enter a number or specify an AWS IoT Events input attribute.

newSimpleRule

Arguments

:: Text
$sel:inputProperty:SimpleRule'

-> ComparisonOperator
$sel:comparisonOperator:SimpleRule'

-> Text
$sel:threshold:SimpleRule'

-> SimpleRule

Create a value of SimpleRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputProperty:SimpleRule', simpleRule_inputProperty - The value on the left side of the comparison operator. You can specify an AWS IoT Events input attribute as an input property.

$sel:comparisonOperator:SimpleRule', simpleRule_comparisonOperator - The comparison operator.

$sel:threshold:SimpleRule', simpleRule_threshold - The value on the right side of the comparison operator. You can enter a number or specify an AWS IoT Events input attribute.

simpleRule_inputProperty :: Lens' SimpleRule Text

The value on the left side of the comparison operator. You can specify an AWS IoT Events input attribute as an input property.

simpleRule_comparisonOperator :: Lens' SimpleRule ComparisonOperator

The comparison operator.

simpleRule_threshold :: Lens' SimpleRule Text

The value on the right side of the comparison operator. You can enter a number or specify an AWS IoT Events input attribute.