Network.AWS.IoTEvents.Types.TransitionEvent

Description

Documentation

data TransitionEvent

Specifies the actions performed and the next state entered when a condition evaluates to TRUE.

See: newTransitionEvent smart constructor.

Constructors

TransitionEvent'

Fields

actions :: Maybe [Action]

The actions to be performed.

eventName :: Text

The name of the transition event.

condition :: Text

Required. A Boolean expression that when TRUE causes the actions to be performed and the nextState to be entered.

nextState :: Text

The next state to enter.

newTransitionEvent

Arguments

:: Text
$sel:eventName:TransitionEvent'

-> Text
$sel:condition:TransitionEvent'

-> Text
$sel:nextState:TransitionEvent'

-> TransitionEvent

Create a value of TransitionEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:actions:TransitionEvent', transitionEvent_actions - The actions to be performed.

$sel:eventName:TransitionEvent', transitionEvent_eventName - The name of the transition event.

$sel:condition:TransitionEvent', transitionEvent_condition - Required. A Boolean expression that when TRUE causes the actions to be performed and the nextState to be entered.

$sel:nextState:TransitionEvent', transitionEvent_nextState - The next state to enter.

transitionEvent_actions :: Lens' TransitionEvent (Maybe [Action])

The actions to be performed.

transitionEvent_eventName :: Lens' TransitionEvent Text

The name of the transition event.

transitionEvent_condition :: Lens' TransitionEvent Text

Required. A Boolean expression that when TRUE causes the actions to be performed and the nextState to be entered.

transitionEvent_nextState :: Lens' TransitionEvent Text

The next state to enter.