Network.AWS.IoTEvents.UpdateDetectorModel

Description

Updates a detector model. Detectors (instances) spawned by the previous version are deleted and then re-created as new inputs arrive.

Creating a Request

data UpdateDetectorModel

See: newUpdateDetectorModel smart constructor.

Constructors

UpdateDetectorModel'

Fields

detectorModelDescription :: Maybe Text

A brief description of the detector model.

evaluationMethod :: Maybe EvaluationMethod

Information about the order in which events are evaluated and how actions are executed.

detectorModelName :: Text

The name of the detector model that is updated.

detectorModelDefinition :: DetectorModelDefinition

Information that defines how a detector operates.

roleArn :: Text

The ARN of the role that grants permission to AWS IoT Events to perform its operations.

newUpdateDetectorModel

Arguments

:: Text
$sel:detectorModelName:UpdateDetectorModel'

-> DetectorModelDefinition
$sel:detectorModelDefinition:UpdateDetectorModel'

-> Text
$sel:roleArn:UpdateDetectorModel'

-> UpdateDetectorModel

Create a value of UpdateDetectorModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:detectorModelDescription:UpdateDetectorModel', updateDetectorModel_detectorModelDescription - A brief description of the detector model.

$sel:evaluationMethod:UpdateDetectorModel', updateDetectorModel_evaluationMethod - Information about the order in which events are evaluated and how actions are executed.

$sel:detectorModelName:UpdateDetectorModel', updateDetectorModel_detectorModelName - The name of the detector model that is updated.

$sel:detectorModelDefinition:UpdateDetectorModel', updateDetectorModel_detectorModelDefinition - Information that defines how a detector operates.

$sel:roleArn:UpdateDetectorModel', updateDetectorModel_roleArn - The ARN of the role that grants permission to AWS IoT Events to perform its operations.

Request Lenses

updateDetectorModel_detectorModelDescription :: Lens' UpdateDetectorModel (Maybe Text)

A brief description of the detector model.

updateDetectorModel_evaluationMethod :: Lens' UpdateDetectorModel (Maybe EvaluationMethod)

Information about the order in which events are evaluated and how actions are executed.

updateDetectorModel_detectorModelName :: Lens' UpdateDetectorModel Text

The name of the detector model that is updated.

updateDetectorModel_detectorModelDefinition :: Lens' UpdateDetectorModel DetectorModelDefinition

Information that defines how a detector operates.

updateDetectorModel_roleArn :: Lens' UpdateDetectorModel Text

The ARN of the role that grants permission to AWS IoT Events to perform its operations.

Destructuring the Response

data UpdateDetectorModelResponse

See: newUpdateDetectorModelResponse smart constructor.

Constructors

UpdateDetectorModelResponse'

Fields

detectorModelConfiguration :: Maybe DetectorModelConfiguration

Information about how the detector model is configured.

httpStatus :: Int

The response's http status code.

newUpdateDetectorModelResponse

Arguments

:: Int
$sel:httpStatus:UpdateDetectorModelResponse'

-> UpdateDetectorModelResponse

Create a value of UpdateDetectorModelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:detectorModelConfiguration:UpdateDetectorModelResponse', updateDetectorModelResponse_detectorModelConfiguration - Information about how the detector model is configured.

$sel:httpStatus:UpdateDetectorModelResponse', updateDetectorModelResponse_httpStatus - The response's http status code.

Response Lenses

updateDetectorModelResponse_detectorModelConfiguration :: Lens' UpdateDetectorModelResponse (Maybe DetectorModelConfiguration)

Information about how the detector model is configured.

updateDetectorModelResponse_httpStatus :: Lens' UpdateDetectorModelResponse Int

The response's http status code.