Network.AWS.IoTEventsData

Description

Derived from API version 2018-10-23 of the AWS service descriptions, licensed under Apache 2.0.

AWS IoT Events monitors your equipment or device fleets for failures or changes in operation, and triggers actions when such events occur. You can use AWS IoT Events Data API commands to send inputs to detectors, list detectors, and view or update a detector's status.

For more information, see What is AWS IoT Events? in the AWS IoT Events Developer Guide.

Service Configuration

defaultService :: Service

API version 2018-10-23 of the Amazon IoT Events Data SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by IoTEventsData.

InvalidRequestException

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request was invalid.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request could not be completed due to throttling.

InternalFailureException

_InternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

An internal failure occurred.

ServiceUnavailableException

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The service is currently unavailable.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The resource was not found.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

BatchSnoozeAlarm

data BatchSnoozeAlarm

See: newBatchSnoozeAlarm smart constructor.

Constructors

BatchSnoozeAlarm' (NonEmpty SnoozeAlarmActionRequest)

newBatchSnoozeAlarm

Arguments

:: NonEmpty SnoozeAlarmActionRequest
$sel:snoozeActionRequests:BatchSnoozeAlarm'

-> BatchSnoozeAlarm

Create a value of BatchSnoozeAlarm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snoozeActionRequests:BatchSnoozeAlarm', batchSnoozeAlarm_snoozeActionRequests - The list of snooze action requests. You can specify up to 10 requests per operation.

data BatchSnoozeAlarmResponse

See: newBatchSnoozeAlarmResponse smart constructor.

Constructors

BatchSnoozeAlarmResponse' (Maybe [BatchAlarmActionErrorEntry]) Int

newBatchSnoozeAlarmResponse

Arguments

:: Int
$sel:httpStatus:BatchSnoozeAlarmResponse'

-> BatchSnoozeAlarmResponse

Create a value of BatchSnoozeAlarmResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorEntries:BatchSnoozeAlarmResponse', batchSnoozeAlarmResponse_errorEntries - A list of errors associated with the request, or null if there are no errors. Each error entry contains an entry ID that helps you identify the entry that failed.

$sel:httpStatus:BatchSnoozeAlarmResponse', batchSnoozeAlarmResponse_httpStatus - The response's http status code.

BatchDisableAlarm

data BatchDisableAlarm

See: newBatchDisableAlarm smart constructor.

Constructors

BatchDisableAlarm' (NonEmpty DisableAlarmActionRequest)

newBatchDisableAlarm

Arguments

:: NonEmpty DisableAlarmActionRequest
$sel:disableActionRequests:BatchDisableAlarm'

-> BatchDisableAlarm

Create a value of BatchDisableAlarm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:disableActionRequests:BatchDisableAlarm', batchDisableAlarm_disableActionRequests - The list of disable action requests. You can specify up to 10 requests per operation.

data BatchDisableAlarmResponse

See: newBatchDisableAlarmResponse smart constructor.

Constructors

BatchDisableAlarmResponse' (Maybe [BatchAlarmActionErrorEntry]) Int

newBatchDisableAlarmResponse

Arguments

:: Int
$sel:httpStatus:BatchDisableAlarmResponse'

-> BatchDisableAlarmResponse

Create a value of BatchDisableAlarmResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorEntries:BatchDisableAlarmResponse', batchDisableAlarmResponse_errorEntries - A list of errors associated with the request, or null if there are no errors. Each error entry contains an entry ID that helps you identify the entry that failed.

$sel:httpStatus:BatchDisableAlarmResponse', batchDisableAlarmResponse_httpStatus - The response's http status code.

DescribeAlarm

data DescribeAlarm

See: newDescribeAlarm smart constructor.

Constructors

DescribeAlarm' (Maybe Text) Text

newDescribeAlarm

Arguments

:: Text
$sel:alarmModelName:DescribeAlarm'

-> DescribeAlarm

Create a value of DescribeAlarm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyValue:DescribeAlarm', describeAlarm_keyValue - The value of the key used as a filter to select only the alarms associated with the key.

$sel:alarmModelName:DescribeAlarm', describeAlarm_alarmModelName - The name of the alarm model.

data DescribeAlarmResponse

See: newDescribeAlarmResponse smart constructor.

Constructors

DescribeAlarmResponse' (Maybe Alarm) Int

newDescribeAlarmResponse

Arguments

:: Int
$sel:httpStatus:DescribeAlarmResponse'

-> DescribeAlarmResponse

Create a value of DescribeAlarmResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alarm:DescribeAlarmResponse', describeAlarmResponse_alarm - Contains information about an alarm.

$sel:httpStatus:DescribeAlarmResponse', describeAlarmResponse_httpStatus - The response's http status code.

BatchPutMessage

data BatchPutMessage

See: newBatchPutMessage smart constructor.

Constructors

BatchPutMessage' (NonEmpty Message)

newBatchPutMessage

Arguments

:: NonEmpty Message
$sel:messages:BatchPutMessage'

-> BatchPutMessage

Create a value of BatchPutMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:messages:BatchPutMessage', batchPutMessage_messages - The list of messages to send. Each message has the following format: '{ "messageId": "string", "inputName": "string", "payload": "string"}'

data BatchPutMessageResponse

See: newBatchPutMessageResponse smart constructor.

Constructors

BatchPutMessageResponse' (Maybe [BatchPutMessageErrorEntry]) Int

newBatchPutMessageResponse

Arguments

:: Int
$sel:httpStatus:BatchPutMessageResponse'

-> BatchPutMessageResponse

Create a value of BatchPutMessageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:batchPutMessageErrorEntries:BatchPutMessageResponse', batchPutMessageResponse_batchPutMessageErrorEntries - A list of any errors encountered when sending the messages.

$sel:httpStatus:BatchPutMessageResponse', batchPutMessageResponse_httpStatus - The response's http status code.

DescribeDetector

data DescribeDetector

See: newDescribeDetector smart constructor.

Constructors

DescribeDetector' (Maybe Text) Text

newDescribeDetector

Arguments

:: Text
$sel:detectorModelName:DescribeDetector'

-> DescribeDetector

Create a value of DescribeDetector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyValue:DescribeDetector', describeDetector_keyValue - A filter used to limit results to detectors (instances) created because of the given key ID.

$sel:detectorModelName:DescribeDetector', describeDetector_detectorModelName - The name of the detector model whose detectors (instances) you want information about.

data DescribeDetectorResponse

See: newDescribeDetectorResponse smart constructor.

Constructors

DescribeDetectorResponse' (Maybe Detector) Int

newDescribeDetectorResponse

Arguments

:: Int
$sel:httpStatus:DescribeDetectorResponse'

-> DescribeDetectorResponse

Create a value of DescribeDetectorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:detector:DescribeDetectorResponse', describeDetectorResponse_detector - Information about the detector (instance).

$sel:httpStatus:DescribeDetectorResponse', describeDetectorResponse_httpStatus - The response's http status code.

BatchUpdateDetector

data BatchUpdateDetector

See: newBatchUpdateDetector smart constructor.

Constructors

BatchUpdateDetector' (NonEmpty UpdateDetectorRequest)

newBatchUpdateDetector

Arguments

:: NonEmpty UpdateDetectorRequest
$sel:detectors:BatchUpdateDetector'

-> BatchUpdateDetector

Create a value of BatchUpdateDetector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:detectors:BatchUpdateDetector', batchUpdateDetector_detectors - The list of detectors (instances) to update, along with the values to update.

data BatchUpdateDetectorResponse

See: newBatchUpdateDetectorResponse smart constructor.

Constructors

BatchUpdateDetectorResponse' (Maybe [BatchUpdateDetectorErrorEntry]) Int

newBatchUpdateDetectorResponse

Arguments

:: Int
$sel:httpStatus:BatchUpdateDetectorResponse'

-> BatchUpdateDetectorResponse

Create a value of BatchUpdateDetectorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:batchUpdateDetectorErrorEntries:BatchUpdateDetectorResponse', batchUpdateDetectorResponse_batchUpdateDetectorErrorEntries - A list of those detector updates that resulted in errors. (If an error is listed here, the specific update did not occur.)

$sel:httpStatus:BatchUpdateDetectorResponse', batchUpdateDetectorResponse_httpStatus - The response's http status code.

BatchAcknowledgeAlarm

data BatchAcknowledgeAlarm

See: newBatchAcknowledgeAlarm smart constructor.

Constructors

BatchAcknowledgeAlarm' (NonEmpty AcknowledgeAlarmActionRequest)

newBatchAcknowledgeAlarm

Arguments

:: NonEmpty AcknowledgeAlarmActionRequest
$sel:acknowledgeActionRequests:BatchAcknowledgeAlarm'

-> BatchAcknowledgeAlarm

Create a value of BatchAcknowledgeAlarm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:acknowledgeActionRequests:BatchAcknowledgeAlarm', batchAcknowledgeAlarm_acknowledgeActionRequests - The list of acknowledge action requests. You can specify up to 10 requests per operation.

data BatchAcknowledgeAlarmResponse

See: newBatchAcknowledgeAlarmResponse smart constructor.

Constructors

BatchAcknowledgeAlarmResponse' (Maybe [BatchAlarmActionErrorEntry]) Int

newBatchAcknowledgeAlarmResponse

Arguments

:: Int
$sel:httpStatus:BatchAcknowledgeAlarmResponse'

-> BatchAcknowledgeAlarmResponse

Create a value of BatchAcknowledgeAlarmResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorEntries:BatchAcknowledgeAlarmResponse', batchAcknowledgeAlarmResponse_errorEntries - A list of errors associated with the request, or null if there are no errors. Each error entry contains an entry ID that helps you identify the entry that failed.

$sel:httpStatus:BatchAcknowledgeAlarmResponse', batchAcknowledgeAlarmResponse_httpStatus - The response's http status code.

ListAlarms

data ListAlarms

See: newListAlarms smart constructor.

Constructors

ListAlarms' (Maybe Text) (Maybe Natural) Text

newListAlarms

Arguments

:: Text
$sel:alarmModelName:ListAlarms'

-> ListAlarms

Create a value of ListAlarms with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAlarms', listAlarms_nextToken - The token that you can use to return the next set of results.

$sel:maxResults:ListAlarms', listAlarms_maxResults - The maximum number of results to be returned per request.

$sel:alarmModelName:ListAlarms', listAlarms_alarmModelName - The name of the alarm model.

data ListAlarmsResponse

See: newListAlarmsResponse smart constructor.

Constructors

ListAlarmsResponse' (Maybe Text) (Maybe [AlarmSummary]) Int

newListAlarmsResponse

Arguments

:: Int
$sel:httpStatus:ListAlarmsResponse'

-> ListAlarmsResponse

Create a value of ListAlarmsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAlarms', listAlarmsResponse_nextToken - The token that you can use to return the next set of results, or null if there are no more results.

$sel:alarmSummaries:ListAlarmsResponse', listAlarmsResponse_alarmSummaries - A list that summarizes each alarm.

$sel:httpStatus:ListAlarmsResponse', listAlarmsResponse_httpStatus - The response's http status code.

BatchResetAlarm

data BatchResetAlarm

See: newBatchResetAlarm smart constructor.

Constructors

BatchResetAlarm' (NonEmpty ResetAlarmActionRequest)

newBatchResetAlarm

Arguments

:: NonEmpty ResetAlarmActionRequest
$sel:resetActionRequests:BatchResetAlarm'

-> BatchResetAlarm

Create a value of BatchResetAlarm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resetActionRequests:BatchResetAlarm', batchResetAlarm_resetActionRequests - The list of reset action requests. You can specify up to 10 requests per operation.

data BatchResetAlarmResponse

See: newBatchResetAlarmResponse smart constructor.

Constructors

BatchResetAlarmResponse' (Maybe [BatchAlarmActionErrorEntry]) Int

newBatchResetAlarmResponse

Arguments

:: Int
$sel:httpStatus:BatchResetAlarmResponse'

-> BatchResetAlarmResponse

Create a value of BatchResetAlarmResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorEntries:BatchResetAlarmResponse', batchResetAlarmResponse_errorEntries - A list of errors associated with the request, or null if there are no errors. Each error entry contains an entry ID that helps you identify the entry that failed.

$sel:httpStatus:BatchResetAlarmResponse', batchResetAlarmResponse_httpStatus - The response's http status code.

ListDetectors

data ListDetectors

See: newListDetectors smart constructor.

Constructors

ListDetectors' (Maybe Text) (Maybe Text) (Maybe Natural) Text

newListDetectors

Arguments

:: Text
$sel:detectorModelName:ListDetectors'

-> ListDetectors

Create a value of ListDetectors with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDetectors', listDetectors_nextToken - The token that you can use to return the next set of results.

$sel:stateName:ListDetectors', listDetectors_stateName - A filter that limits results to those detectors (instances) in the given state.

$sel:maxResults:ListDetectors', listDetectors_maxResults - The maximum number of results to be returned per request.

$sel:detectorModelName:ListDetectors', listDetectors_detectorModelName - The name of the detector model whose detectors (instances) are listed.

data ListDetectorsResponse

See: newListDetectorsResponse smart constructor.

Constructors

ListDetectorsResponse' (Maybe Text) (Maybe [DetectorSummary]) Int

newListDetectorsResponse

Arguments

:: Int
$sel:httpStatus:ListDetectorsResponse'

-> ListDetectorsResponse

Create a value of ListDetectorsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDetectors', listDetectorsResponse_nextToken - The token that you can use to return the next set of results, or null if there are no more results.

$sel:detectorSummaries:ListDetectorsResponse', listDetectorsResponse_detectorSummaries - A list of summary information about the detectors (instances).

$sel:httpStatus:ListDetectorsResponse', listDetectorsResponse_httpStatus - The response's http status code.

BatchEnableAlarm

data BatchEnableAlarm

See: newBatchEnableAlarm smart constructor.

Constructors

BatchEnableAlarm' (NonEmpty EnableAlarmActionRequest)

newBatchEnableAlarm

Arguments

:: NonEmpty EnableAlarmActionRequest
$sel:enableActionRequests:BatchEnableAlarm'

-> BatchEnableAlarm

Create a value of BatchEnableAlarm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enableActionRequests:BatchEnableAlarm', batchEnableAlarm_enableActionRequests - The list of enable action requests. You can specify up to 10 requests per operation.

data BatchEnableAlarmResponse

See: newBatchEnableAlarmResponse smart constructor.

Constructors

BatchEnableAlarmResponse' (Maybe [BatchAlarmActionErrorEntry]) Int

newBatchEnableAlarmResponse

Arguments

:: Int
$sel:httpStatus:BatchEnableAlarmResponse'

-> BatchEnableAlarmResponse

Create a value of BatchEnableAlarmResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorEntries:BatchEnableAlarmResponse', batchEnableAlarmResponse_errorEntries - A list of errors associated with the request, or null if there are no errors. Each error entry contains an entry ID that helps you identify the entry that failed.

$sel:httpStatus:BatchEnableAlarmResponse', batchEnableAlarmResponse_httpStatus - The response's http status code.

Types

AlarmStateName

newtype AlarmStateName

Constructors

AlarmStateName'

Fields

fromAlarmStateName :: Text

Bundled Patterns

pattern AlarmStateName_ACKNOWLEDGED :: AlarmStateName

pattern AlarmStateName_ACTIVE :: AlarmStateName

pattern AlarmStateName_DISABLED :: AlarmStateName

pattern AlarmStateName_LATCHED :: AlarmStateName

pattern AlarmStateName_NORMAL :: AlarmStateName

pattern AlarmStateName_SNOOZE_DISABLED :: AlarmStateName

ComparisonOperator

newtype ComparisonOperator

Constructors

ComparisonOperator'

Fields

fromComparisonOperator :: Text

Bundled Patterns

pattern ComparisonOperator_EQUAL :: ComparisonOperator

pattern ComparisonOperator_GREATER :: ComparisonOperator

pattern ComparisonOperator_GREATER_OR_EQUAL :: ComparisonOperator

pattern ComparisonOperator_LESS :: ComparisonOperator

pattern ComparisonOperator_LESS_OR_EQUAL :: ComparisonOperator

pattern ComparisonOperator_NOT_EQUAL :: ComparisonOperator

CustomerActionName

newtype CustomerActionName

Constructors

CustomerActionName'

Fields

fromCustomerActionName :: Text

Bundled Patterns

pattern CustomerActionName_ACKNOWLEDGE :: CustomerActionName

pattern CustomerActionName_DISABLE :: CustomerActionName

pattern CustomerActionName_ENABLE :: CustomerActionName

pattern CustomerActionName_RESET :: CustomerActionName

pattern CustomerActionName_SNOOZE :: CustomerActionName

ErrorCode

newtype ErrorCode

Constructors

ErrorCode'

Fields

fromErrorCode :: Text

Bundled Patterns

pattern ErrorCode_InternalFailureException :: ErrorCode

pattern ErrorCode_InvalidRequestException :: ErrorCode

pattern ErrorCode_ResourceNotFoundException :: ErrorCode

pattern ErrorCode_ServiceUnavailableException :: ErrorCode

pattern ErrorCode_ThrottlingException :: ErrorCode

EventType

newtype EventType

Constructors

EventType'

Fields

fromEventType :: Text

Bundled Patterns

pattern EventType_STATE_CHANGE :: EventType

TriggerType

newtype TriggerType

Constructors

TriggerType'

Fields

fromTriggerType :: Text

Bundled Patterns

pattern TriggerType_SNOOZE_TIMEOUT :: TriggerType

AcknowledgeActionConfiguration

data AcknowledgeActionConfiguration

Contains the configuration information of an acknowledge action.

See: newAcknowledgeActionConfiguration smart constructor.

Constructors

AcknowledgeActionConfiguration' (Maybe Text)

newAcknowledgeActionConfiguration :: AcknowledgeActionConfiguration

Create a value of AcknowledgeActionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:note:AcknowledgeActionConfiguration', acknowledgeActionConfiguration_note - The note that you can leave when you acknowledge the alarm.

AcknowledgeAlarmActionRequest

data AcknowledgeAlarmActionRequest

Information needed to acknowledge the alarm.

See: newAcknowledgeAlarmActionRequest smart constructor.

Constructors

AcknowledgeAlarmActionRequest' (Maybe Text) (Maybe Text) Text Text

newAcknowledgeAlarmActionRequest

Arguments

:: Text
$sel:requestId:AcknowledgeAlarmActionRequest'

-> Text
$sel:alarmModelName:AcknowledgeAlarmActionRequest'

-> AcknowledgeAlarmActionRequest

Create a value of AcknowledgeAlarmActionRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyValue:AcknowledgeAlarmActionRequest', acknowledgeAlarmActionRequest_keyValue - The value of the key used as a filter to select only the alarms associated with the key.

$sel:note:AcknowledgeAlarmActionRequest', acknowledgeAlarmActionRequest_note - The note that you can leave when you acknowledge the alarm.

$sel:requestId:AcknowledgeAlarmActionRequest', acknowledgeAlarmActionRequest_requestId - The request ID. Each ID must be unique within each batch.

$sel:alarmModelName:AcknowledgeAlarmActionRequest', acknowledgeAlarmActionRequest_alarmModelName - The name of the alarm model.

Alarm

data Alarm

Contains information about an alarm.

See: newAlarm smart constructor.

Constructors

Alarm' (Maybe Text) (Maybe POSIX) (Maybe AlarmState) (Maybe Text) (Maybe Natural) (Maybe POSIX) (Maybe Text)

newAlarm :: Alarm

Create a value of Alarm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyValue:Alarm', alarm_keyValue - The value of the key used as a filter to select only the alarms associated with the key.

$sel:creationTime:Alarm', alarm_creationTime - The time the alarm was created, in the Unix epoch format.

$sel:alarmState:Alarm', alarm_alarmState - Contains information about the current state of the alarm.

$sel:alarmModelName:Alarm', alarm_alarmModelName - The name of the alarm model.

$sel:severity:Alarm', alarm_severity - A non-negative integer that reflects the severity level of the alarm.

$sel:lastUpdateTime:Alarm', alarm_lastUpdateTime - The time the alarm was last updated, in the Unix epoch format.

$sel:alarmModelVersion:Alarm', alarm_alarmModelVersion - The version of the alarm model.

AlarmState

data AlarmState

Contains information about the current state of the alarm.

See: newAlarmState smart constructor.

Constructors

AlarmState' (Maybe CustomerAction) (Maybe AlarmStateName) (Maybe RuleEvaluation) (Maybe SystemEvent)

newAlarmState :: AlarmState

Create a value of AlarmState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customerAction:AlarmState', alarmState_customerAction - Contains information about the action that you can take to respond to the alarm.

$sel:stateName:AlarmState', alarmState_stateName - The name of the alarm state. The state name can be one of the following values:

DISABLED - When the alarm is in the DISABLED state, it isn't ready to evaluate data. To enable the alarm, you must change the alarm to the NORMAL state.

NORMAL - When the alarm is in the NORMAL state, it's ready to evaluate data.

ACTIVE - If the alarm is in the ACTIVE state, the alarm is invoked.

ACKNOWLEDGED - When the alarm is in the ACKNOWLEDGED state, the alarm was invoked and you acknowledged the alarm.

SNOOZE_DISABLED - When the alarm is in the SNOOZE_DISABLED state, the alarm is disabled for a specified period of time. After the snooze time, the alarm automatically changes to the NORMAL state.

LATCHED - When the alarm is in the LATCHED state, the alarm was invoked. However, the data that the alarm is currently evaluating is within the specified range. To change the alarm to the NORMAL state, you must acknowledge the alarm.

$sel:ruleEvaluation:AlarmState', alarmState_ruleEvaluation - Information needed to evaluate data.

$sel:systemEvent:AlarmState', alarmState_systemEvent - Contains information about alarm state changes.

AlarmSummary

data AlarmSummary

Contains a summary of an alarm.

See: newAlarmSummary smart constructor.

Constructors

AlarmSummary' (Maybe Text) (Maybe POSIX) (Maybe Text) (Maybe AlarmStateName) (Maybe POSIX) (Maybe Text)

newAlarmSummary :: AlarmSummary

Create a value of AlarmSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyValue:AlarmSummary', alarmSummary_keyValue - The value of the key used as a filter to select only the alarms associated with the key.

$sel:creationTime:AlarmSummary', alarmSummary_creationTime - The time the alarm was created, in the Unix epoch format.

$sel:alarmModelName:AlarmSummary', alarmSummary_alarmModelName - The name of the alarm model.

$sel:stateName:AlarmSummary', alarmSummary_stateName - The name of the alarm state. The state name can be one of the following values:

DISABLED - When the alarm is in the DISABLED state, it isn't ready to evaluate data. To enable the alarm, you must change the alarm to the NORMAL state.

NORMAL - When the alarm is in the NORMAL state, it's ready to evaluate data.

ACTIVE - If the alarm is in the ACTIVE state, the alarm is invoked.

ACKNOWLEDGED - When the alarm is in the ACKNOWLEDGED state, the alarm was invoked and you acknowledged the alarm.

SNOOZE_DISABLED - When the alarm is in the SNOOZE_DISABLED state, the alarm is disabled for a specified period of time. After the snooze time, the alarm automatically changes to the NORMAL state.

LATCHED - When the alarm is in the LATCHED state, the alarm was invoked. However, the data that the alarm is currently evaluating is within the specified range. To change the alarm to the NORMAL state, you must acknowledge the alarm.

$sel:lastUpdateTime:AlarmSummary', alarmSummary_lastUpdateTime - The time the alarm was last updated, in the Unix epoch format.

$sel:alarmModelVersion:AlarmSummary', alarmSummary_alarmModelVersion - The version of the alarm model.

BatchAlarmActionErrorEntry

data BatchAlarmActionErrorEntry

Contains error messages associated with one of the following requests:

BatchAcknowledgeAlarm

BatchDisableAlarm

BatchEnableAlarm

BatchResetAlarm

BatchSnoozeAlarm

See: newBatchAlarmActionErrorEntry smart constructor.

Constructors

BatchAlarmActionErrorEntry' (Maybe Text) (Maybe ErrorCode) (Maybe Text)

newBatchAlarmActionErrorEntry :: BatchAlarmActionErrorEntry

Create a value of BatchAlarmActionErrorEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:BatchAlarmActionErrorEntry', batchAlarmActionErrorEntry_requestId - The request ID. Each ID must be unique within each batch.

$sel:errorCode:BatchAlarmActionErrorEntry', batchAlarmActionErrorEntry_errorCode - The error code.

$sel:errorMessage:BatchAlarmActionErrorEntry', batchAlarmActionErrorEntry_errorMessage - A message that describes the error.

BatchPutMessageErrorEntry

data BatchPutMessageErrorEntry

Contains information about the errors encountered.

See: newBatchPutMessageErrorEntry smart constructor.

Constructors

BatchPutMessageErrorEntry' (Maybe ErrorCode) (Maybe Text) (Maybe Text)

newBatchPutMessageErrorEntry :: BatchPutMessageErrorEntry

Create a value of BatchPutMessageErrorEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorCode:BatchPutMessageErrorEntry', batchPutMessageErrorEntry_errorCode - The error code.

$sel:errorMessage:BatchPutMessageErrorEntry', batchPutMessageErrorEntry_errorMessage - A message that describes the error.

$sel:messageId:BatchPutMessageErrorEntry', batchPutMessageErrorEntry_messageId - The ID of the message that caused the error. (See the value corresponding to the "messageId" key in the "message" object.)

BatchUpdateDetectorErrorEntry

data BatchUpdateDetectorErrorEntry

Information about the error that occurred when attempting to update a detector.

See: newBatchUpdateDetectorErrorEntry smart constructor.

Constructors

BatchUpdateDetectorErrorEntry' (Maybe ErrorCode) (Maybe Text) (Maybe Text)

newBatchUpdateDetectorErrorEntry :: BatchUpdateDetectorErrorEntry

Create a value of BatchUpdateDetectorErrorEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorCode:BatchUpdateDetectorErrorEntry', batchUpdateDetectorErrorEntry_errorCode - The error code.

$sel:errorMessage:BatchUpdateDetectorErrorEntry', batchUpdateDetectorErrorEntry_errorMessage - A message that describes the error.

$sel:messageId:BatchUpdateDetectorErrorEntry', batchUpdateDetectorErrorEntry_messageId - The "messageId" of the update request that caused the error. (The value of the "messageId" in the update request "Detector" object.)

CustomerAction

data CustomerAction

Contains information about the action that you can take to respond to the alarm.

See: newCustomerAction smart constructor.

Constructors

CustomerAction' (Maybe ResetActionConfiguration) (Maybe CustomerActionName) (Maybe SnoozeActionConfiguration) (Maybe EnableActionConfiguration) (Maybe DisableActionConfiguration) (Maybe AcknowledgeActionConfiguration)

newCustomerAction :: CustomerAction

Create a value of CustomerAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resetActionConfiguration:CustomerAction', customerAction_resetActionConfiguration - Contains the configuration information of a reset action.

$sel:actionName:CustomerAction', customerAction_actionName - The name of the action. The action name can be one of the following values:

SNOOZE - When you snooze the alarm, the alarm state changes to SNOOZE_DISABLED.

ENABLE - When you enable the alarm, the alarm state changes to NORMAL.

DISABLE - When you disable the alarm, the alarm state changes to DISABLED.

ACKNOWLEDGE - When you acknowledge the alarm, the alarm state changes to ACKNOWLEDGED.

RESET - When you reset the alarm, the alarm state changes to NORMAL.

For more information, see the AlarmState API.

$sel:snoozeActionConfiguration:CustomerAction', customerAction_snoozeActionConfiguration - Contains the configuration information of a snooze action.

$sel:enableActionConfiguration:CustomerAction', customerAction_enableActionConfiguration - Contains the configuration information of an enable action.

$sel:disableActionConfiguration:CustomerAction', customerAction_disableActionConfiguration - Contains the configuration information of a disable action.

$sel:acknowledgeActionConfiguration:CustomerAction', customerAction_acknowledgeActionConfiguration - Contains the configuration information of an acknowledge action.

Detector

data Detector

Information about the detector (instance).

See: newDetector smart constructor.

Constructors

Detector' (Maybe Text) (Maybe POSIX) (Maybe DetectorState) (Maybe Text) (Maybe Text) (Maybe POSIX)

newDetector :: Detector

Create a value of Detector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyValue:Detector', detector_keyValue - The value of the key (identifying the device or system) that caused the creation of this detector (instance).

$sel:creationTime:Detector', detector_creationTime - The time the detector (instance) was created.

$sel:state:Detector', detector_state - The current state of the detector (instance).

$sel:detectorModelName:Detector', detector_detectorModelName - The name of the detector model that created this detector (instance).

$sel:detectorModelVersion:Detector', detector_detectorModelVersion - The version of the detector model that created this detector (instance).

$sel:lastUpdateTime:Detector', detector_lastUpdateTime - The time the detector (instance) was last updated.

DetectorState

data DetectorState

Information about the current state of the detector instance.

See: newDetectorState smart constructor.

Constructors

DetectorState' Text [Variable] [Timer]

newDetectorState

Arguments

:: Text
$sel:stateName:DetectorState'

-> DetectorState

Create a value of DetectorState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stateName:DetectorState', detectorState_stateName - The name of the state.

$sel:variables:DetectorState', detectorState_variables - The current values of the detector's variables.

$sel:timers:DetectorState', detectorState_timers - The current state of the detector's timers.

DetectorStateDefinition

data DetectorStateDefinition

The new state, variable values, and timer settings of the detector (instance).

See: newDetectorStateDefinition smart constructor.

Constructors

DetectorStateDefinition' Text [VariableDefinition] [TimerDefinition]

newDetectorStateDefinition

Arguments

:: Text
$sel:stateName:DetectorStateDefinition'

-> DetectorStateDefinition

Create a value of DetectorStateDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stateName:DetectorStateDefinition', detectorStateDefinition_stateName - The name of the new state of the detector (instance).

$sel:variables:DetectorStateDefinition', detectorStateDefinition_variables - The new values of the detector's variables. Any variable whose value isn't specified is cleared.

$sel:timers:DetectorStateDefinition', detectorStateDefinition_timers - The new values of the detector's timers. Any timer whose value isn't specified is cleared, and its timeout event won't occur.

DetectorStateSummary

data DetectorStateSummary

Information about the detector state.

See: newDetectorStateSummary smart constructor.

Constructors

DetectorStateSummary' (Maybe Text)

newDetectorStateSummary :: DetectorStateSummary

Create a value of DetectorStateSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stateName:DetectorStateSummary', detectorStateSummary_stateName - The name of the state.

DetectorSummary

data DetectorSummary

Information about the detector (instance).

See: newDetectorSummary smart constructor.

Constructors

DetectorSummary' (Maybe Text) (Maybe POSIX) (Maybe DetectorStateSummary) (Maybe Text) (Maybe Text) (Maybe POSIX)

newDetectorSummary :: DetectorSummary

Create a value of DetectorSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyValue:DetectorSummary', detectorSummary_keyValue - The value of the key (identifying the device or system) that caused the creation of this detector (instance).

$sel:creationTime:DetectorSummary', detectorSummary_creationTime - The time the detector (instance) was created.

$sel:state:DetectorSummary', detectorSummary_state - The current state of the detector (instance).

$sel:detectorModelName:DetectorSummary', detectorSummary_detectorModelName - The name of the detector model that created this detector (instance).

$sel:detectorModelVersion:DetectorSummary', detectorSummary_detectorModelVersion - The version of the detector model that created this detector (instance).

$sel:lastUpdateTime:DetectorSummary', detectorSummary_lastUpdateTime - The time the detector (instance) was last updated.

DisableActionConfiguration

data DisableActionConfiguration

Contains the configuration information of a disable action.

See: newDisableActionConfiguration smart constructor.

Constructors

DisableActionConfiguration' (Maybe Text)

newDisableActionConfiguration :: DisableActionConfiguration

Create a value of DisableActionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:note:DisableActionConfiguration', disableActionConfiguration_note - The note that you can leave when you disable the alarm.

DisableAlarmActionRequest

data DisableAlarmActionRequest

Information used to disable the alarm.

See: newDisableAlarmActionRequest smart constructor.

Constructors

DisableAlarmActionRequest' (Maybe Text) (Maybe Text) Text Text

newDisableAlarmActionRequest

Arguments

:: Text
$sel:requestId:DisableAlarmActionRequest'

-> Text
$sel:alarmModelName:DisableAlarmActionRequest'

-> DisableAlarmActionRequest

Create a value of DisableAlarmActionRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyValue:DisableAlarmActionRequest', disableAlarmActionRequest_keyValue - The value of the key used as a filter to select only the alarms associated with the key.

$sel:note:DisableAlarmActionRequest', disableAlarmActionRequest_note - The note that you can leave when you disable the alarm.

$sel:requestId:DisableAlarmActionRequest', disableAlarmActionRequest_requestId - The request ID. Each ID must be unique within each batch.

$sel:alarmModelName:DisableAlarmActionRequest', disableAlarmActionRequest_alarmModelName - The name of the alarm model.

EnableActionConfiguration

data EnableActionConfiguration

Contains the configuration information of an enable action.

See: newEnableActionConfiguration smart constructor.

Constructors

EnableActionConfiguration' (Maybe Text)

newEnableActionConfiguration :: EnableActionConfiguration

Create a value of EnableActionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:note:EnableActionConfiguration', enableActionConfiguration_note - The note that you can leave when you enable the alarm.

EnableAlarmActionRequest

data EnableAlarmActionRequest

Information needed to enable the alarm.

See: newEnableAlarmActionRequest smart constructor.

Constructors

EnableAlarmActionRequest' (Maybe Text) (Maybe Text) Text Text

newEnableAlarmActionRequest

Arguments

:: Text
$sel:requestId:EnableAlarmActionRequest'

-> Text
$sel:alarmModelName:EnableAlarmActionRequest'

-> EnableAlarmActionRequest

Create a value of EnableAlarmActionRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyValue:EnableAlarmActionRequest', enableAlarmActionRequest_keyValue - The value of the key used as a filter to select only the alarms associated with the key.

$sel:note:EnableAlarmActionRequest', enableAlarmActionRequest_note - The note that you can leave when you enable the alarm.

$sel:requestId:EnableAlarmActionRequest', enableAlarmActionRequest_requestId - The request ID. Each ID must be unique within each batch.

$sel:alarmModelName:EnableAlarmActionRequest', enableAlarmActionRequest_alarmModelName - The name of the alarm model.

Message

data Message

Information about a message.

See: newMessage smart constructor.

Constructors

Message' (Maybe TimestampValue) Text Text Base64

newMessage

Arguments

:: Text
$sel:messageId:Message'

-> Text
$sel:inputName:Message'

-> ByteString
$sel:payload:Message'

-> Message

Create a value of Message with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timestamp:Message', message_timestamp - The timestamp associated with the message.

$sel:messageId:Message', message_messageId - The ID to assign to the message. Within each batch sent, each "messageId" must be unique.

$sel:inputName:Message', message_inputName - The name of the input into which the message payload is transformed.

$sel:payload:Message', message_payload - The payload of the message. This can be a JSON string or a Base-64-encoded string representing binary data (in which case you must decode it).-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

ResetActionConfiguration

data ResetActionConfiguration

Contains the configuration information of a reset action.

See: newResetActionConfiguration smart constructor.

Constructors

ResetActionConfiguration' (Maybe Text)

newResetActionConfiguration :: ResetActionConfiguration

Create a value of ResetActionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:note:ResetActionConfiguration', resetActionConfiguration_note - The note that you can leave when you reset the alarm.

ResetAlarmActionRequest

data ResetAlarmActionRequest

Information needed to reset the alarm.

See: newResetAlarmActionRequest smart constructor.

Constructors

ResetAlarmActionRequest' (Maybe Text) (Maybe Text) Text Text

newResetAlarmActionRequest

Arguments

:: Text
$sel:requestId:ResetAlarmActionRequest'

-> Text
$sel:alarmModelName:ResetAlarmActionRequest'

-> ResetAlarmActionRequest

Create a value of ResetAlarmActionRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyValue:ResetAlarmActionRequest', resetAlarmActionRequest_keyValue - The value of the key used as a filter to select only the alarms associated with the key.

$sel:note:ResetAlarmActionRequest', resetAlarmActionRequest_note - The note that you can leave when you reset the alarm.

$sel:requestId:ResetAlarmActionRequest', resetAlarmActionRequest_requestId - The request ID. Each ID must be unique within each batch.

$sel:alarmModelName:ResetAlarmActionRequest', resetAlarmActionRequest_alarmModelName - The name of the alarm model.

RuleEvaluation

data RuleEvaluation

Information needed to evaluate data.

See: newRuleEvaluation smart constructor.

Constructors

RuleEvaluation' (Maybe SimpleRuleEvaluation)

newRuleEvaluation :: RuleEvaluation

Create a value of RuleEvaluation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:simpleRuleEvaluation:RuleEvaluation', ruleEvaluation_simpleRuleEvaluation - Information needed to compare two values with a comparison operator.

SimpleRuleEvaluation

data SimpleRuleEvaluation

Information needed to compare two values with a comparison operator.

See: newSimpleRuleEvaluation smart constructor.

Constructors

SimpleRuleEvaluation' (Maybe Text) (Maybe Text) (Maybe ComparisonOperator)

newSimpleRuleEvaluation :: SimpleRuleEvaluation

Create a value of SimpleRuleEvaluation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thresholdValue:SimpleRuleEvaluation', simpleRuleEvaluation_thresholdValue - The threshold value, on the right side of the comparison operator.

$sel:inputPropertyValue:SimpleRuleEvaluation', simpleRuleEvaluation_inputPropertyValue - The value of the input property, on the left side of the comparison operator.

$sel:operator:SimpleRuleEvaluation', simpleRuleEvaluation_operator - The comparison operator.

SnoozeActionConfiguration

data SnoozeActionConfiguration

Contains the configuration information of a snooze action.

See: newSnoozeActionConfiguration smart constructor.

Constructors

SnoozeActionConfiguration' (Maybe Text) (Maybe Int)

newSnoozeActionConfiguration :: SnoozeActionConfiguration

Create a value of SnoozeActionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:note:SnoozeActionConfiguration', snoozeActionConfiguration_note - The note that you can leave when you snooze the alarm.

$sel:snoozeDuration:SnoozeActionConfiguration', snoozeActionConfiguration_snoozeDuration - The snooze time in seconds. The alarm automatically changes to the NORMAL state after this duration.

SnoozeAlarmActionRequest

data SnoozeAlarmActionRequest

Information needed to snooze the alarm.

See: newSnoozeAlarmActionRequest smart constructor.

Constructors

SnoozeAlarmActionRequest' (Maybe Text) (Maybe Text) Text Text Int

newSnoozeAlarmActionRequest

Arguments

:: Text
$sel:requestId:SnoozeAlarmActionRequest'

-> Text
$sel:alarmModelName:SnoozeAlarmActionRequest'

-> Int
$sel:snoozeDuration:SnoozeAlarmActionRequest'

-> SnoozeAlarmActionRequest

Create a value of SnoozeAlarmActionRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyValue:SnoozeAlarmActionRequest', snoozeAlarmActionRequest_keyValue - The value of the key used as a filter to select only the alarms associated with the key.

$sel:note:SnoozeAlarmActionRequest', snoozeAlarmActionRequest_note - The note that you can leave when you snooze the alarm.

$sel:requestId:SnoozeAlarmActionRequest', snoozeAlarmActionRequest_requestId - The request ID. Each ID must be unique within each batch.

$sel:alarmModelName:SnoozeAlarmActionRequest', snoozeAlarmActionRequest_alarmModelName - The name of the alarm model.

$sel:snoozeDuration:SnoozeAlarmActionRequest', snoozeAlarmActionRequest_snoozeDuration - The snooze time in seconds. The alarm automatically changes to the NORMAL state after this duration.

StateChangeConfiguration

data StateChangeConfiguration

Contains the configuration information of alarm state changes.

See: newStateChangeConfiguration smart constructor.

Constructors

StateChangeConfiguration' (Maybe TriggerType)

newStateChangeConfiguration :: StateChangeConfiguration

Create a value of StateChangeConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:triggerType:StateChangeConfiguration', stateChangeConfiguration_triggerType - The trigger type. If the value is SNOOZE_TIMEOUT, the snooze duration ends and the alarm automatically changes to the NORMAL state.

SystemEvent

data SystemEvent

Contains information about alarm state changes.

See: newSystemEvent smart constructor.

Constructors

SystemEvent' (Maybe EventType) (Maybe StateChangeConfiguration)

newSystemEvent :: SystemEvent

Create a value of SystemEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventType:SystemEvent', systemEvent_eventType - The event type. If the value is STATE_CHANGE, the event contains information about alarm state changes.

$sel:stateChangeConfiguration:SystemEvent', systemEvent_stateChangeConfiguration - Contains the configuration information of alarm state changes.

Timer

data Timer

The current state of a timer.

See: newTimer smart constructor.

Constructors

Timer' Text POSIX

newTimer

Arguments

:: Text
$sel:name:Timer'

-> UTCTime
$sel:timestamp:Timer'

-> Timer

Create a value of Timer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:Timer', timer_name - The name of the timer.

$sel:timestamp:Timer', timer_timestamp - The number of seconds which have elapsed on the timer.

TimerDefinition

data TimerDefinition

The new setting of a timer.

See: newTimerDefinition smart constructor.

Constructors

TimerDefinition' Text Int

newTimerDefinition

Arguments

:: Text
$sel:name:TimerDefinition'

-> Int
$sel:seconds:TimerDefinition'

-> TimerDefinition

Create a value of TimerDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:TimerDefinition', timerDefinition_name - The name of the timer.

$sel:seconds:TimerDefinition', timerDefinition_seconds - The new setting of the timer (the number of seconds before the timer elapses).

TimestampValue

data TimestampValue

Contains information about a timestamp.

See: newTimestampValue smart constructor.

Constructors

TimestampValue' (Maybe Natural)

newTimestampValue :: TimestampValue

Create a value of TimestampValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timeInMillis:TimestampValue', timestampValue_timeInMillis - The value of the timestamp, in the Unix epoch format.

UpdateDetectorRequest

data UpdateDetectorRequest

Information used to update the detector (instance).

See: newUpdateDetectorRequest smart constructor.

Constructors

UpdateDetectorRequest' (Maybe Text) Text Text DetectorStateDefinition

newUpdateDetectorRequest

Arguments

:: Text
$sel:messageId:UpdateDetectorRequest'

-> Text
$sel:detectorModelName:UpdateDetectorRequest'

-> DetectorStateDefinition
$sel:state:UpdateDetectorRequest'

-> UpdateDetectorRequest

Create a value of UpdateDetectorRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyValue:UpdateDetectorRequest', updateDetectorRequest_keyValue - The value of the input key attribute (identifying the device or system) that caused the creation of this detector (instance).

$sel:messageId:UpdateDetectorRequest', updateDetectorRequest_messageId - The ID to assign to the detector update "message". Each "messageId" must be unique within each batch sent.

$sel:detectorModelName:UpdateDetectorRequest', updateDetectorRequest_detectorModelName - The name of the detector model that created the detectors (instances).

$sel:state:UpdateDetectorRequest', updateDetectorRequest_state - The new state, variable values, and timer settings of the detector (instance).

Variable

data Variable

The current state of the variable.

See: newVariable smart constructor.

Constructors

Variable' Text Text

newVariable

Arguments

:: Text
$sel:name:Variable'

-> Text
$sel:value:Variable'

-> Variable

Create a value of Variable with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:Variable', variable_name - The name of the variable.

$sel:value:Variable', variable_value - The current value of the variable.

VariableDefinition

data VariableDefinition

The new value of the variable.

See: newVariableDefinition smart constructor.

Constructors

VariableDefinition' Text Text

newVariableDefinition

Arguments

:: Text
$sel:name:VariableDefinition'

-> Text
$sel:value:VariableDefinition'

-> VariableDefinition

Create a value of VariableDefinition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:VariableDefinition', variableDefinition_name - The name of the variable.

$sel:value:VariableDefinition', variableDefinition_value - The new value of the variable.