Network.AWS.IoTEventsData.BatchDisableAlarm

Description

Disables one or more alarms. The alarms change to the DISABLED state after you disable them.

Creating a Request

data BatchDisableAlarm

See: newBatchDisableAlarm smart constructor.

Constructors

BatchDisableAlarm'

Fields

disableActionRequests :: NonEmpty DisableAlarmActionRequest

The list of disable action requests. You can specify up to 10 requests per operation.

newBatchDisableAlarm

Arguments

:: NonEmpty DisableAlarmActionRequest
$sel:disableActionRequests:BatchDisableAlarm'

-> BatchDisableAlarm

Create a value of BatchDisableAlarm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:disableActionRequests:BatchDisableAlarm', batchDisableAlarm_disableActionRequests - The list of disable action requests. You can specify up to 10 requests per operation.

Request Lenses

batchDisableAlarm_disableActionRequests :: Lens' BatchDisableAlarm (NonEmpty DisableAlarmActionRequest)

The list of disable action requests. You can specify up to 10 requests per operation.

Destructuring the Response

data BatchDisableAlarmResponse

See: newBatchDisableAlarmResponse smart constructor.

Constructors

BatchDisableAlarmResponse'

Fields

errorEntries :: Maybe [BatchAlarmActionErrorEntry]

A list of errors associated with the request, or null if there are no errors. Each error entry contains an entry ID that helps you identify the entry that failed.

httpStatus :: Int

The response's http status code.

newBatchDisableAlarmResponse

Arguments

:: Int
$sel:httpStatus:BatchDisableAlarmResponse'

-> BatchDisableAlarmResponse

Create a value of BatchDisableAlarmResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorEntries:BatchDisableAlarmResponse', batchDisableAlarmResponse_errorEntries - A list of errors associated with the request, or null if there are no errors. Each error entry contains an entry ID that helps you identify the entry that failed.

$sel:httpStatus:BatchDisableAlarmResponse', batchDisableAlarmResponse_httpStatus - The response's http status code.

Response Lenses

batchDisableAlarmResponse_errorEntries :: Lens' BatchDisableAlarmResponse (Maybe [BatchAlarmActionErrorEntry])

A list of errors associated with the request, or null if there are no errors. Each error entry contains an entry ID that helps you identify the entry that failed.

batchDisableAlarmResponse_httpStatus :: Lens' BatchDisableAlarmResponse Int

The response's http status code.