Network.AWS.IoTEventsData.Types.Alarm

Description

Documentation

data Alarm

Contains information about an alarm.

See: newAlarm smart constructor.

Constructors

Alarm'

Fields

keyValue :: Maybe Text

The value of the key used as a filter to select only the alarms associated with the key.

creationTime :: Maybe POSIX

The time the alarm was created, in the Unix epoch format.

alarmState :: Maybe AlarmState

Contains information about the current state of the alarm.

alarmModelName :: Maybe Text

The name of the alarm model.

severity :: Maybe Natural

A non-negative integer that reflects the severity level of the alarm.

lastUpdateTime :: Maybe POSIX

The time the alarm was last updated, in the Unix epoch format.

alarmModelVersion :: Maybe Text

The version of the alarm model.

newAlarm :: Alarm

Create a value of Alarm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyValue:Alarm', alarm_keyValue - The value of the key used as a filter to select only the alarms associated with the key.

$sel:creationTime:Alarm', alarm_creationTime - The time the alarm was created, in the Unix epoch format.

$sel:alarmState:Alarm', alarm_alarmState - Contains information about the current state of the alarm.

$sel:alarmModelName:Alarm', alarm_alarmModelName - The name of the alarm model.

$sel:severity:Alarm', alarm_severity - A non-negative integer that reflects the severity level of the alarm.

$sel:lastUpdateTime:Alarm', alarm_lastUpdateTime - The time the alarm was last updated, in the Unix epoch format.

$sel:alarmModelVersion:Alarm', alarm_alarmModelVersion - The version of the alarm model.

alarm_keyValue :: Lens' Alarm (Maybe Text)

The value of the key used as a filter to select only the alarms associated with the key.

alarm_creationTime :: Lens' Alarm (Maybe UTCTime)

The time the alarm was created, in the Unix epoch format.

alarm_alarmState :: Lens' Alarm (Maybe AlarmState)

Contains information about the current state of the alarm.

alarm_alarmModelName :: Lens' Alarm (Maybe Text)

The name of the alarm model.

alarm_severity :: Lens' Alarm (Maybe Natural)

A non-negative integer that reflects the severity level of the alarm.

alarm_lastUpdateTime :: Lens' Alarm (Maybe UTCTime)

The time the alarm was last updated, in the Unix epoch format.

alarm_alarmModelVersion :: Lens' Alarm (Maybe Text)

The version of the alarm model.