Network.AWS.IoTEventsData.Types.BatchUpdateDetectorErrorEntry

Description

Documentation

data BatchUpdateDetectorErrorEntry

Information about the error that occurred when attempting to update a detector.

See: newBatchUpdateDetectorErrorEntry smart constructor.

Constructors

BatchUpdateDetectorErrorEntry'

Fields

errorCode :: Maybe ErrorCode

The error code.

errorMessage :: Maybe Text

A message that describes the error.

messageId :: Maybe Text

The "messageId" of the update request that caused the error. (The value of the "messageId" in the update request "Detector" object.)

newBatchUpdateDetectorErrorEntry :: BatchUpdateDetectorErrorEntry

Create a value of BatchUpdateDetectorErrorEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorCode:BatchUpdateDetectorErrorEntry', batchUpdateDetectorErrorEntry_errorCode - The error code.

$sel:errorMessage:BatchUpdateDetectorErrorEntry', batchUpdateDetectorErrorEntry_errorMessage - A message that describes the error.

$sel:messageId:BatchUpdateDetectorErrorEntry', batchUpdateDetectorErrorEntry_messageId - The "messageId" of the update request that caused the error. (The value of the "messageId" in the update request "Detector" object.)

batchUpdateDetectorErrorEntry_errorCode :: Lens' BatchUpdateDetectorErrorEntry (Maybe ErrorCode)

The error code.

batchUpdateDetectorErrorEntry_errorMessage :: Lens' BatchUpdateDetectorErrorEntry (Maybe Text)

A message that describes the error.

batchUpdateDetectorErrorEntry_messageId :: Lens' BatchUpdateDetectorErrorEntry (Maybe Text)

The "messageId" of the update request that caused the error. (The value of the "messageId" in the update request "Detector" object.)