Network.AWS.IoTFleetHub

Description

Derived from API version 2020-11-03 of the AWS service descriptions, licensed under Apache 2.0.

With Fleet Hub for AWS IoT Device Management you can build stand-alone web applications for monitoring the health of your device fleets.

Fleet Hub for AWS IoT Device Management is in public preview and is subject to change.

Service Configuration

defaultService :: Service

API version 2020-11-03 of the Amazon IoT Fleet Hub SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by IoTFleetHub.

InvalidRequestException

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request is not valid.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The request conflicts with the current state of the resource.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The rate exceeds the limit.

InternalFailureException

_InternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

An unexpected error has occurred.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource does not exist.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

A limit has been exceeded.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

DescribeApplication

data DescribeApplication

See: newDescribeApplication smart constructor.

Constructors

DescribeApplication' Text

newDescribeApplication

Arguments

:: Text
$sel:applicationId:DescribeApplication'

-> DescribeApplication

Create a value of DescribeApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:DescribeApplication', describeApplication_applicationId - The unique Id of the web application.

data DescribeApplicationResponse

See: newDescribeApplicationResponse smart constructor.

Constructors

DescribeApplicationResponse' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text)) Int Text Text Text Text ApplicationState Integer Integer Text

newDescribeApplicationResponse

Arguments

:: Int
$sel:httpStatus:DescribeApplicationResponse'

-> Text
$sel:applicationId:DescribeApplication'

-> Text
$sel:applicationArn:DescribeApplicationResponse'

-> Text
$sel:applicationName:DescribeApplicationResponse'

-> Text
$sel:applicationUrl:DescribeApplicationResponse'

-> ApplicationState
$sel:applicationState:DescribeApplicationResponse'

-> Integer
$sel:applicationCreationDate:DescribeApplicationResponse'

-> Integer
$sel:applicationLastUpdateDate:DescribeApplicationResponse'

-> Text
$sel:roleArn:DescribeApplicationResponse'

-> DescribeApplicationResponse

Create a value of DescribeApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationDescription:DescribeApplicationResponse', describeApplicationResponse_applicationDescription - An optional description of the web application.

$sel:ssoClientId:DescribeApplicationResponse', describeApplicationResponse_ssoClientId - The Id of the single sign-on client that you use to authenticate and authorize users on the web application.

$sel:errorMessage:DescribeApplicationResponse', describeApplicationResponse_errorMessage - A message indicating why the DescribeApplication API failed.

$sel:tags:DescribeApplicationResponse', describeApplicationResponse_tags - A set of key/value pairs that you can use to manage the web application resource.

$sel:httpStatus:DescribeApplicationResponse', describeApplicationResponse_httpStatus - The response's http status code.

$sel:applicationId:DescribeApplication', describeApplicationResponse_applicationId - The unique Id of the web application.

$sel:applicationArn:DescribeApplicationResponse', describeApplicationResponse_applicationArn - The ARN of the web application.

$sel:applicationName:DescribeApplicationResponse', describeApplicationResponse_applicationName - The name of the web application.

$sel:applicationUrl:DescribeApplicationResponse', describeApplicationResponse_applicationUrl - The URL of the web application.

$sel:applicationState:DescribeApplicationResponse', describeApplicationResponse_applicationState - The current state of the web application.

$sel:applicationCreationDate:DescribeApplicationResponse', describeApplicationResponse_applicationCreationDate - The date (in Unix epoch time) when the application was created.

$sel:applicationLastUpdateDate:DescribeApplicationResponse', describeApplicationResponse_applicationLastUpdateDate - The date (in Unix epoch time) when the application was last updated.

$sel:roleArn:DescribeApplicationResponse', describeApplicationResponse_roleArn - The ARN of the role that the web application assumes when it interacts with AWS IoT Core.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The ARN of the resource.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (HashMap Text Text)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - The list of tags assigned to the resource.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

DeleteApplication

data DeleteApplication

See: newDeleteApplication smart constructor.

Constructors

DeleteApplication' (Maybe Text) Text

newDeleteApplication

Arguments

:: Text
$sel:applicationId:DeleteApplication'

-> DeleteApplication

Create a value of DeleteApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:DeleteApplication', deleteApplication_clientToken - A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

$sel:applicationId:DeleteApplication', deleteApplication_applicationId - The unique Id of the web application.

data DeleteApplicationResponse

See: newDeleteApplicationResponse smart constructor.

Constructors

DeleteApplicationResponse' Int

newDeleteApplicationResponse

Arguments

:: Int
$sel:httpStatus:DeleteApplicationResponse'

-> DeleteApplicationResponse

Create a value of DeleteApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteApplicationResponse', deleteApplicationResponse_httpStatus - The response's http status code.

UpdateApplication

data UpdateApplication

See: newUpdateApplication smart constructor.

Constructors

UpdateApplication' (Maybe Text) (Maybe Text) (Maybe Text) Text

newUpdateApplication

Arguments

:: Text
$sel:applicationId:UpdateApplication'

-> UpdateApplication

Create a value of UpdateApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationDescription:UpdateApplication', updateApplication_applicationDescription - An optional description of the web application.

$sel:clientToken:UpdateApplication', updateApplication_clientToken - A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

$sel:applicationName:UpdateApplication', updateApplication_applicationName - The name of the web application.

$sel:applicationId:UpdateApplication', updateApplication_applicationId - The unique Id of the web application.

data UpdateApplicationResponse

See: newUpdateApplicationResponse smart constructor.

Constructors

UpdateApplicationResponse' Int

newUpdateApplicationResponse

Arguments

:: Int
$sel:httpStatus:UpdateApplicationResponse'

-> UpdateApplicationResponse

Create a value of UpdateApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateApplicationResponse', updateApplicationResponse_httpStatus - The response's http status code.

CreateApplication

data CreateApplication

See: newCreateApplication smart constructor.

Constructors

CreateApplication' (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text)) Text Text

newCreateApplication

Arguments

:: Text
$sel:applicationName:CreateApplication'

-> Text
$sel:roleArn:CreateApplication'

-> CreateApplication

Create a value of CreateApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationDescription:CreateApplication', createApplication_applicationDescription - An optional description of the web application.

$sel:clientToken:CreateApplication', createApplication_clientToken - A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

$sel:tags:CreateApplication', createApplication_tags - A set of key/value pairs that you can use to manage the web application resource.

$sel:applicationName:CreateApplication', createApplication_applicationName - The name of the web application.

$sel:roleArn:CreateApplication', createApplication_roleArn - The ARN of the role that the web application assumes when it interacts with AWS IoT Core.

The name of the role must be in the form AWSIotFleetHub_random_string .

data CreateApplicationResponse

See: newCreateApplicationResponse smart constructor.

Constructors

CreateApplicationResponse' Int Text Text

newCreateApplicationResponse

Arguments

:: Int
$sel:httpStatus:CreateApplicationResponse'

-> Text
$sel:applicationId:CreateApplicationResponse'

-> Text
$sel:applicationArn:CreateApplicationResponse'

-> CreateApplicationResponse

Create a value of CreateApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateApplicationResponse', createApplicationResponse_httpStatus - The response's http status code.

$sel:applicationId:CreateApplicationResponse', createApplicationResponse_applicationId - The unique Id of the web application.

$sel:applicationArn:CreateApplicationResponse', createApplicationResponse_applicationArn - The ARN of the web application.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (HashMap Text Text)

newTagResource

Arguments

:: Text
$sel:resourceArn:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:TagResource', tagResource_resourceArn - The ARN of the resource.

$sel:tags:TagResource', tagResource_tags - The new or modified tags for the resource.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

ListApplications (Paginated)

data ListApplications

See: newListApplications smart constructor.

Constructors

ListApplications' (Maybe Text)

newListApplications :: ListApplications

Create a value of ListApplications with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListApplications', listApplications_nextToken - A token used to get the next set of results.

data ListApplicationsResponse

See: newListApplicationsResponse smart constructor.

Constructors

ListApplicationsResponse' (Maybe Text) (Maybe [ApplicationSummary]) Int

newListApplicationsResponse

Arguments

:: Int
$sel:httpStatus:ListApplicationsResponse'

-> ListApplicationsResponse

Create a value of ListApplicationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListApplications', listApplicationsResponse_nextToken - A token used to get the next set of results.

$sel:applicationSummaries:ListApplicationsResponse', listApplicationsResponse_applicationSummaries - An array of objects that provide summaries of information about the web applications in the list.

$sel:httpStatus:ListApplicationsResponse', listApplicationsResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text [Text]

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:UntagResource', untagResource_resourceArn - The ARN of the resource.

$sel:tagKeys:UntagResource', untagResource_tagKeys - A list of the keys of the tags to be removed from the resource.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

Types

ApplicationState

newtype ApplicationState

Constructors

ApplicationState'

Fields

fromApplicationState :: Text

Bundled Patterns

pattern ApplicationState_ACTIVE :: ApplicationState

pattern ApplicationState_CREATE_FAILED :: ApplicationState

pattern ApplicationState_CREATING :: ApplicationState

pattern ApplicationState_DELETE_FAILED :: ApplicationState

pattern ApplicationState_DELETING :: ApplicationState

ApplicationSummary

data ApplicationSummary

A summary of information about a AWS IoT Device Management web application.

Fleet Hub for AWS IoT Device Management is in public preview and is subject to change.

See: newApplicationSummary smart constructor.

Constructors

ApplicationSummary' (Maybe Text) (Maybe ApplicationState) (Maybe Integer) (Maybe Integer) Text Text Text

newApplicationSummary

Arguments

:: Text
$sel:applicationId:ApplicationSummary'

-> Text
$sel:applicationName:ApplicationSummary'

-> Text
$sel:applicationUrl:ApplicationSummary'

-> ApplicationSummary

Create a value of ApplicationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationDescription:ApplicationSummary', applicationSummary_applicationDescription - An optional description of the web application.

$sel:applicationState:ApplicationSummary', applicationSummary_applicationState - The current state of the web application.

$sel:applicationCreationDate:ApplicationSummary', applicationSummary_applicationCreationDate - The date (in Unix epoch time) when the web application was created.

$sel:applicationLastUpdateDate:ApplicationSummary', applicationSummary_applicationLastUpdateDate - The date (in Unix epoch time) when the web application was last updated.

$sel:applicationId:ApplicationSummary', applicationSummary_applicationId - The unique Id of the web application.

$sel:applicationName:ApplicationSummary', applicationSummary_applicationName - The name of the web application.

$sel:applicationUrl:ApplicationSummary', applicationSummary_applicationUrl - The URL of the web application.