Network.AWS.IoTFleetHub.Types

Description

Service Configuration

defaultService :: Service

API version 2020-11-03 of the Amazon IoT Fleet Hub SDK configuration.

Errors

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request is not valid.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

The request conflicts with the current state of the resource.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The rate exceeds the limit.

_InternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

An unexpected error has occurred.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource does not exist.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

A limit has been exceeded.

ApplicationState

newtype ApplicationState

Constructors

ApplicationState'

Fields

fromApplicationState :: Text

Bundled Patterns

pattern ApplicationState_ACTIVE :: ApplicationState

pattern ApplicationState_CREATE_FAILED :: ApplicationState

pattern ApplicationState_CREATING :: ApplicationState

pattern ApplicationState_DELETE_FAILED :: ApplicationState

pattern ApplicationState_DELETING :: ApplicationState

ApplicationSummary

data ApplicationSummary

A summary of information about a AWS IoT Device Management web application.

Fleet Hub for AWS IoT Device Management is in public preview and is subject to change.

See: newApplicationSummary smart constructor.

Constructors

ApplicationSummary'

Fields

applicationDescription :: Maybe Text

An optional description of the web application.

applicationState :: Maybe ApplicationState

The current state of the web application.

applicationCreationDate :: Maybe Integer

The date (in Unix epoch time) when the web application was created.

applicationLastUpdateDate :: Maybe Integer

The date (in Unix epoch time) when the web application was last updated.

applicationId :: Text

The unique Id of the web application.

applicationName :: Text

The name of the web application.

applicationUrl :: Text

The URL of the web application.

newApplicationSummary

Arguments

:: Text
$sel:applicationId:ApplicationSummary'

-> Text
$sel:applicationName:ApplicationSummary'

-> Text
$sel:applicationUrl:ApplicationSummary'

-> ApplicationSummary

Create a value of ApplicationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationDescription:ApplicationSummary', applicationSummary_applicationDescription - An optional description of the web application.

$sel:applicationState:ApplicationSummary', applicationSummary_applicationState - The current state of the web application.

$sel:applicationCreationDate:ApplicationSummary', applicationSummary_applicationCreationDate - The date (in Unix epoch time) when the web application was created.

$sel:applicationLastUpdateDate:ApplicationSummary', applicationSummary_applicationLastUpdateDate - The date (in Unix epoch time) when the web application was last updated.

$sel:applicationId:ApplicationSummary', applicationSummary_applicationId - The unique Id of the web application.

$sel:applicationName:ApplicationSummary', applicationSummary_applicationName - The name of the web application.

$sel:applicationUrl:ApplicationSummary', applicationSummary_applicationUrl - The URL of the web application.

applicationSummary_applicationDescription :: Lens' ApplicationSummary (Maybe Text)

An optional description of the web application.

applicationSummary_applicationState :: Lens' ApplicationSummary (Maybe ApplicationState)

The current state of the web application.

applicationSummary_applicationCreationDate :: Lens' ApplicationSummary (Maybe Integer)

The date (in Unix epoch time) when the web application was created.

applicationSummary_applicationLastUpdateDate :: Lens' ApplicationSummary (Maybe Integer)

The date (in Unix epoch time) when the web application was last updated.

applicationSummary_applicationId :: Lens' ApplicationSummary Text

The unique Id of the web application.

applicationSummary_applicationName :: Lens' ApplicationSummary Text

The name of the web application.

applicationSummary_applicationUrl :: Lens' ApplicationSummary Text

The URL of the web application.