Network.AWS.IoTJobsData

Description

Derived from API version 2017-09-29 of the AWS service descriptions, licensed under Apache 2.0.

AWS IoT Jobs is a service that allows you to define a set of jobs — remote operations that are sent to and executed on one or more devices connected to AWS IoT. For example, you can define a job that instructs a set of devices to download and install application or firmware updates, reboot, rotate certificates, or perform remote troubleshooting operations.

To create a job, you make a job document which is a description of the remote operations to be performed, and you specify a list of targets that should perform the operations. The targets can be individual things, thing groups or both.

AWS IoT Jobs sends a message to inform the targets that a job is available. The target starts the execution of the job by downloading the job document, performing the operations it specifies, and reporting its progress to AWS IoT. The Jobs service provides commands to track the progress of a job on a specific target and for all the targets of the job

Service Configuration

defaultService :: Service

API version 2017-09-29 of the Amazon IoT Jobs Data Plane SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by IoTJobsData.

TerminalStateException

_TerminalStateException :: AsError a => Getting (First ServiceError) a ServiceError

The job is in a terminal state.

CertificateValidationException

_CertificateValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The certificate is invalid.

ServiceUnavailableException

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The service is temporarily unavailable.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The rate exceeds the limit.

InvalidRequestException

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The contents of the request were invalid. For example, this code is returned when an UpdateJobExecution request contains invalid status details. The message contains details about the error.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource does not exist.

InvalidStateTransitionException

_InvalidStateTransitionException :: AsError a => Getting (First ServiceError) a ServiceError

An update attempted to change the job execution to a state that is invalid because of the job execution's current state (for example, an attempt to change a request in state SUCCESS to state IN_PROGRESS). In this case, the body of the error message also contains the executionState field.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

GetPendingJobExecutions

data GetPendingJobExecutions

See: newGetPendingJobExecutions smart constructor.

Constructors

GetPendingJobExecutions' Text

newGetPendingJobExecutions

Arguments

:: Text
$sel:thingName:GetPendingJobExecutions'

-> GetPendingJobExecutions

Create a value of GetPendingJobExecutions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thingName:GetPendingJobExecutions', getPendingJobExecutions_thingName - The name of the thing that is executing the job.

data GetPendingJobExecutionsResponse

See: newGetPendingJobExecutionsResponse smart constructor.

Constructors

GetPendingJobExecutionsResponse' (Maybe [JobExecutionSummary]) (Maybe [JobExecutionSummary]) Int

newGetPendingJobExecutionsResponse

Arguments

:: Int
$sel:httpStatus:GetPendingJobExecutionsResponse'

-> GetPendingJobExecutionsResponse

Create a value of GetPendingJobExecutionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inProgressJobs:GetPendingJobExecutionsResponse', getPendingJobExecutionsResponse_inProgressJobs - A list of JobExecutionSummary objects with status IN_PROGRESS.

$sel:queuedJobs:GetPendingJobExecutionsResponse', getPendingJobExecutionsResponse_queuedJobs - A list of JobExecutionSummary objects with status QUEUED.

$sel:httpStatus:GetPendingJobExecutionsResponse', getPendingJobExecutionsResponse_httpStatus - The response's http status code.

UpdateJobExecution

data UpdateJobExecution

See: newUpdateJobExecution smart constructor.

Constructors

UpdateJobExecution' (Maybe Bool) (Maybe Integer) (Maybe Bool) (Maybe (HashMap Text Text)) (Maybe Integer) (Maybe Integer) Text Text JobExecutionStatus

newUpdateJobExecution

Arguments

:: Text
$sel:jobId:UpdateJobExecution'

-> Text
$sel:thingName:UpdateJobExecution'

-> JobExecutionStatus
$sel:status:UpdateJobExecution'

-> UpdateJobExecution

Create a value of UpdateJobExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeJobDocument:UpdateJobExecution', updateJobExecution_includeJobDocument - Optional. When set to true, the response contains the job document. The default is false.

$sel:expectedVersion:UpdateJobExecution', updateJobExecution_expectedVersion - Optional. The expected current version of the job execution. Each time you update the job execution, its version is incremented. If the version of the job execution stored in Jobs does not match, the update is rejected with a VersionMismatch error, and an ErrorResponse that contains the current job execution status data is returned. (This makes it unnecessary to perform a separate DescribeJobExecution request in order to obtain the job execution status data.)

$sel:includeJobExecutionState:UpdateJobExecution', updateJobExecution_includeJobExecutionState - Optional. When included and set to true, the response contains the JobExecutionState data. The default is false.

$sel:statusDetails:UpdateJobExecution', updateJobExecution_statusDetails - Optional. A collection of name/value pairs that describe the status of the job execution. If not specified, the statusDetails are unchanged.

$sel:stepTimeoutInMinutes:UpdateJobExecution', updateJobExecution_stepTimeoutInMinutes - Specifies the amount of time this device has to finish execution of this job. If the job execution status is not set to a terminal state before this timer expires, or before the timer is reset (by again calling UpdateJobExecution, setting the status to IN_PROGRESS and specifying a new timeout value in this field) the job execution status will be automatically set to TIMED_OUT. Note that setting or resetting this timeout has no effect on that job execution timeout which may have been specified when the job was created (CreateJob using field timeoutConfig).

$sel:executionNumber:UpdateJobExecution', updateJobExecution_executionNumber - Optional. A number that identifies a particular job execution on a particular device.

$sel:jobId:UpdateJobExecution', updateJobExecution_jobId - The unique identifier assigned to this job when it was created.

$sel:thingName:UpdateJobExecution', updateJobExecution_thingName - The name of the thing associated with the device.

$sel:status:UpdateJobExecution', updateJobExecution_status - The new status for the job execution (IN_PROGRESS, FAILED, SUCCESS, or REJECTED). This must be specified on every update.

data UpdateJobExecutionResponse

See: newUpdateJobExecutionResponse smart constructor.

Constructors

UpdateJobExecutionResponse' (Maybe JobExecutionState) (Maybe Text) Int

newUpdateJobExecutionResponse

Arguments

:: Int
$sel:httpStatus:UpdateJobExecutionResponse'

-> UpdateJobExecutionResponse

Create a value of UpdateJobExecutionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:executionState:UpdateJobExecutionResponse', updateJobExecutionResponse_executionState - A JobExecutionState object.

$sel:jobDocument:UpdateJobExecutionResponse', updateJobExecutionResponse_jobDocument - The contents of the Job Documents.

$sel:httpStatus:UpdateJobExecutionResponse', updateJobExecutionResponse_httpStatus - The response's http status code.

DescribeJobExecution

data DescribeJobExecution

See: newDescribeJobExecution smart constructor.

Constructors

DescribeJobExecution' (Maybe Bool) (Maybe Integer) Text Text

newDescribeJobExecution

Arguments

:: Text
$sel:jobId:DescribeJobExecution'

-> Text
$sel:thingName:DescribeJobExecution'

-> DescribeJobExecution

Create a value of DescribeJobExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeJobDocument:DescribeJobExecution', describeJobExecution_includeJobDocument - Optional. When set to true, the response contains the job document. The default is false.

$sel:executionNumber:DescribeJobExecution', describeJobExecution_executionNumber - Optional. A number that identifies a particular job execution on a particular device. If not specified, the latest job execution is returned.

$sel:jobId:DescribeJobExecution', describeJobExecution_jobId - The unique identifier assigned to this job when it was created.

$sel:thingName:DescribeJobExecution', describeJobExecution_thingName - The thing name associated with the device the job execution is running on.

data DescribeJobExecutionResponse

See: newDescribeJobExecutionResponse smart constructor.

Constructors

DescribeJobExecutionResponse' (Maybe JobExecution) Int

newDescribeJobExecutionResponse

Arguments

:: Int
$sel:httpStatus:DescribeJobExecutionResponse'

-> DescribeJobExecutionResponse

Create a value of DescribeJobExecutionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:execution:DescribeJobExecutionResponse', describeJobExecutionResponse_execution - Contains data about a job execution.

$sel:httpStatus:DescribeJobExecutionResponse', describeJobExecutionResponse_httpStatus - The response's http status code.

StartNextPendingJobExecution

data StartNextPendingJobExecution

See: newStartNextPendingJobExecution smart constructor.

Constructors

StartNextPendingJobExecution' (Maybe (HashMap Text Text)) (Maybe Integer) Text

newStartNextPendingJobExecution

Arguments

:: Text
$sel:thingName:StartNextPendingJobExecution'

-> StartNextPendingJobExecution

Create a value of StartNextPendingJobExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusDetails:StartNextPendingJobExecution', startNextPendingJobExecution_statusDetails - A collection of name/value pairs that describe the status of the job execution. If not specified, the statusDetails are unchanged.

$sel:stepTimeoutInMinutes:StartNextPendingJobExecution', startNextPendingJobExecution_stepTimeoutInMinutes - Specifies the amount of time this device has to finish execution of this job. If the job execution status is not set to a terminal state before this timer expires, or before the timer is reset (by calling UpdateJobExecution, setting the status to IN_PROGRESS and specifying a new timeout value in field stepTimeoutInMinutes) the job execution status will be automatically set to TIMED_OUT. Note that setting this timeout has no effect on that job execution timeout which may have been specified when the job was created (CreateJob using field timeoutConfig).

$sel:thingName:StartNextPendingJobExecution', startNextPendingJobExecution_thingName - The name of the thing associated with the device.

data StartNextPendingJobExecutionResponse

See: newStartNextPendingJobExecutionResponse smart constructor.

Constructors

StartNextPendingJobExecutionResponse' (Maybe JobExecution) Int

newStartNextPendingJobExecutionResponse

Arguments

:: Int
$sel:httpStatus:StartNextPendingJobExecutionResponse'

-> StartNextPendingJobExecutionResponse

Create a value of StartNextPendingJobExecutionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:execution:StartNextPendingJobExecutionResponse', startNextPendingJobExecutionResponse_execution - A JobExecution object.

$sel:httpStatus:StartNextPendingJobExecutionResponse', startNextPendingJobExecutionResponse_httpStatus - The response's http status code.

Types

JobExecutionStatus

newtype JobExecutionStatus

Constructors

JobExecutionStatus'

Fields

fromJobExecutionStatus :: Text

Bundled Patterns

pattern JobExecutionStatus_CANCELED :: JobExecutionStatus

pattern JobExecutionStatus_FAILED :: JobExecutionStatus

pattern JobExecutionStatus_IN_PROGRESS :: JobExecutionStatus

pattern JobExecutionStatus_QUEUED :: JobExecutionStatus

pattern JobExecutionStatus_REJECTED :: JobExecutionStatus

pattern JobExecutionStatus_REMOVED :: JobExecutionStatus

pattern JobExecutionStatus_SUCCEEDED :: JobExecutionStatus

pattern JobExecutionStatus_TIMED_OUT :: JobExecutionStatus

JobExecution

data JobExecution

Contains data about a job execution.

See: newJobExecution smart constructor.

Constructors

JobExecution' (Maybe Integer) (Maybe JobExecutionStatus) (Maybe (HashMap Text Text)) (Maybe Text) (Maybe Integer) (Maybe Integer) (Maybe Integer) (Maybe Text) (Maybe Integer) (Maybe Text) (Maybe Integer)

newJobExecution :: JobExecution

Create a value of JobExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startedAt:JobExecution', jobExecution_startedAt - The time, in milliseconds since the epoch, when the job execution was started.

$sel:status:JobExecution', jobExecution_status - The status of the job execution. Can be one of: "QUEUED", "IN_PROGRESS", "FAILED", "SUCCESS", "CANCELED", "REJECTED", or "REMOVED".

$sel:statusDetails:JobExecution', jobExecution_statusDetails - A collection of name/value pairs that describe the status of the job execution.

$sel:thingName:JobExecution', jobExecution_thingName - The name of the thing that is executing the job.

$sel:queuedAt:JobExecution', jobExecution_queuedAt - The time, in milliseconds since the epoch, when the job execution was enqueued.

$sel:executionNumber:JobExecution', jobExecution_executionNumber - A number that identifies a particular job execution on a particular device. It can be used later in commands that return or update job execution information.

$sel:versionNumber:JobExecution', jobExecution_versionNumber - The version of the job execution. Job execution versions are incremented each time they are updated by a device.

$sel:jobDocument:JobExecution', jobExecution_jobDocument - The content of the job document.

$sel:approximateSecondsBeforeTimedOut:JobExecution', jobExecution_approximateSecondsBeforeTimedOut - The estimated number of seconds that remain before the job execution status will be changed to TIMED_OUT.

$sel:jobId:JobExecution', jobExecution_jobId - The unique identifier you assigned to this job when it was created.

$sel:lastUpdatedAt:JobExecution', jobExecution_lastUpdatedAt - The time, in milliseconds since the epoch, when the job execution was last updated.

JobExecutionState

data JobExecutionState

Contains data about the state of a job execution.

See: newJobExecutionState smart constructor.

Constructors

JobExecutionState' (Maybe JobExecutionStatus) (Maybe (HashMap Text Text)) (Maybe Integer)

newJobExecutionState :: JobExecutionState

Create a value of JobExecutionState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:JobExecutionState', jobExecutionState_status - The status of the job execution. Can be one of: "QUEUED", "IN_PROGRESS", "FAILED", "SUCCESS", "CANCELED", "REJECTED", or "REMOVED".

$sel:statusDetails:JobExecutionState', jobExecutionState_statusDetails - A collection of name/value pairs that describe the status of the job execution.

$sel:versionNumber:JobExecutionState', jobExecutionState_versionNumber - The version of the job execution. Job execution versions are incremented each time they are updated by a device.

JobExecutionSummary

data JobExecutionSummary

Contains a subset of information about a job execution.

See: newJobExecutionSummary smart constructor.

Constructors

JobExecutionSummary' (Maybe Integer) (Maybe Integer) (Maybe Integer) (Maybe Integer) (Maybe Text) (Maybe Integer)

newJobExecutionSummary :: JobExecutionSummary

Create a value of JobExecutionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startedAt:JobExecutionSummary', jobExecutionSummary_startedAt - The time, in milliseconds since the epoch, when the job execution started.

$sel:queuedAt:JobExecutionSummary', jobExecutionSummary_queuedAt - The time, in milliseconds since the epoch, when the job execution was enqueued.

$sel:executionNumber:JobExecutionSummary', jobExecutionSummary_executionNumber - A number that identifies a particular job execution on a particular device.

$sel:versionNumber:JobExecutionSummary', jobExecutionSummary_versionNumber - The version of the job execution. Job execution versions are incremented each time AWS IoT Jobs receives an update from a device.

$sel:jobId:JobExecutionSummary', jobExecutionSummary_jobId - The unique identifier you assigned to this job when it was created.

$sel:lastUpdatedAt:JobExecutionSummary', jobExecutionSummary_lastUpdatedAt - The time, in milliseconds since the epoch, when the job execution was last updated.