Network.AWS.IoTJobsData.StartNextPendingJobExecution

Description

Gets and starts the next pending (status IN_PROGRESS or QUEUED) job execution for a thing.

Creating a Request

data StartNextPendingJobExecution

See: newStartNextPendingJobExecution smart constructor.

Constructors

StartNextPendingJobExecution'

Fields

statusDetails :: Maybe (HashMap Text Text)

A collection of name/value pairs that describe the status of the job execution. If not specified, the statusDetails are unchanged.

stepTimeoutInMinutes :: Maybe Integer

Specifies the amount of time this device has to finish execution of this job. If the job execution status is not set to a terminal state before this timer expires, or before the timer is reset (by calling UpdateJobExecution, setting the status to IN_PROGRESS and specifying a new timeout value in field stepTimeoutInMinutes) the job execution status will be automatically set to TIMED_OUT. Note that setting this timeout has no effect on that job execution timeout which may have been specified when the job was created (CreateJob using field timeoutConfig).

thingName :: Text

The name of the thing associated with the device.

newStartNextPendingJobExecution

Arguments

:: Text
$sel:thingName:StartNextPendingJobExecution'

-> StartNextPendingJobExecution

Create a value of StartNextPendingJobExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusDetails:StartNextPendingJobExecution', startNextPendingJobExecution_statusDetails - A collection of name/value pairs that describe the status of the job execution. If not specified, the statusDetails are unchanged.

$sel:stepTimeoutInMinutes:StartNextPendingJobExecution', startNextPendingJobExecution_stepTimeoutInMinutes - Specifies the amount of time this device has to finish execution of this job. If the job execution status is not set to a terminal state before this timer expires, or before the timer is reset (by calling UpdateJobExecution, setting the status to IN_PROGRESS and specifying a new timeout value in field stepTimeoutInMinutes) the job execution status will be automatically set to TIMED_OUT. Note that setting this timeout has no effect on that job execution timeout which may have been specified when the job was created (CreateJob using field timeoutConfig).

$sel:thingName:StartNextPendingJobExecution', startNextPendingJobExecution_thingName - The name of the thing associated with the device.

Request Lenses

startNextPendingJobExecution_statusDetails :: Lens' StartNextPendingJobExecution (Maybe (HashMap Text Text))

A collection of name/value pairs that describe the status of the job execution. If not specified, the statusDetails are unchanged.

startNextPendingJobExecution_stepTimeoutInMinutes :: Lens' StartNextPendingJobExecution (Maybe Integer)

Specifies the amount of time this device has to finish execution of this job. If the job execution status is not set to a terminal state before this timer expires, or before the timer is reset (by calling UpdateJobExecution, setting the status to IN_PROGRESS and specifying a new timeout value in field stepTimeoutInMinutes) the job execution status will be automatically set to TIMED_OUT. Note that setting this timeout has no effect on that job execution timeout which may have been specified when the job was created (CreateJob using field timeoutConfig).

startNextPendingJobExecution_thingName :: Lens' StartNextPendingJobExecution Text

The name of the thing associated with the device.

Destructuring the Response

data StartNextPendingJobExecutionResponse

See: newStartNextPendingJobExecutionResponse smart constructor.

Constructors

StartNextPendingJobExecutionResponse'

Fields

execution :: Maybe JobExecution

A JobExecution object.

httpStatus :: Int

The response's http status code.

newStartNextPendingJobExecutionResponse

Arguments

:: Int
$sel:httpStatus:StartNextPendingJobExecutionResponse'

-> StartNextPendingJobExecutionResponse

Create a value of StartNextPendingJobExecutionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:execution:StartNextPendingJobExecutionResponse', startNextPendingJobExecutionResponse_execution - A JobExecution object.

$sel:httpStatus:StartNextPendingJobExecutionResponse', startNextPendingJobExecutionResponse_httpStatus - The response's http status code.

Response Lenses

startNextPendingJobExecutionResponse_execution :: Lens' StartNextPendingJobExecutionResponse (Maybe JobExecution)

A JobExecution object.

startNextPendingJobExecutionResponse_httpStatus :: Lens' StartNextPendingJobExecutionResponse Int

The response's http status code.