Network.AWS.IoTJobsData.Types

Description

Service Configuration

defaultService :: Service

API version 2017-09-29 of the Amazon IoT Jobs Data Plane SDK configuration.

Errors

_TerminalStateException :: AsError a => Getting (First ServiceError) a ServiceError

The job is in a terminal state.

_CertificateValidationException :: AsError a => Getting (First ServiceError) a ServiceError

The certificate is invalid.

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The service is temporarily unavailable.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The rate exceeds the limit.

_InvalidRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The contents of the request were invalid. For example, this code is returned when an UpdateJobExecution request contains invalid status details. The message contains details about the error.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource does not exist.

_InvalidStateTransitionException :: AsError a => Getting (First ServiceError) a ServiceError

An update attempted to change the job execution to a state that is invalid because of the job execution's current state (for example, an attempt to change a request in state SUCCESS to state IN_PROGRESS). In this case, the body of the error message also contains the executionState field.

JobExecutionStatus

newtype JobExecutionStatus

Constructors

JobExecutionStatus'

Fields

fromJobExecutionStatus :: Text

Bundled Patterns

pattern JobExecutionStatus_CANCELED :: JobExecutionStatus

pattern JobExecutionStatus_FAILED :: JobExecutionStatus

pattern JobExecutionStatus_IN_PROGRESS :: JobExecutionStatus

pattern JobExecutionStatus_QUEUED :: JobExecutionStatus

pattern JobExecutionStatus_REJECTED :: JobExecutionStatus

pattern JobExecutionStatus_REMOVED :: JobExecutionStatus

pattern JobExecutionStatus_SUCCEEDED :: JobExecutionStatus

pattern JobExecutionStatus_TIMED_OUT :: JobExecutionStatus

JobExecution

data JobExecution

Contains data about a job execution.

See: newJobExecution smart constructor.

Constructors

JobExecution'

Fields

startedAt :: Maybe Integer

The time, in milliseconds since the epoch, when the job execution was started.

status :: Maybe JobExecutionStatus

The status of the job execution. Can be one of: "QUEUED", "IN_PROGRESS", "FAILED", "SUCCESS", "CANCELED", "REJECTED", or "REMOVED".

statusDetails :: Maybe (HashMap Text Text)

A collection of name/value pairs that describe the status of the job execution.

thingName :: Maybe Text

The name of the thing that is executing the job.

queuedAt :: Maybe Integer

The time, in milliseconds since the epoch, when the job execution was enqueued.

executionNumber :: Maybe Integer

A number that identifies a particular job execution on a particular device. It can be used later in commands that return or update job execution information.

versionNumber :: Maybe Integer

The version of the job execution. Job execution versions are incremented each time they are updated by a device.

jobDocument :: Maybe Text

The content of the job document.

approximateSecondsBeforeTimedOut :: Maybe Integer

The estimated number of seconds that remain before the job execution status will be changed to TIMED_OUT.

jobId :: Maybe Text

The unique identifier you assigned to this job when it was created.

lastUpdatedAt :: Maybe Integer

The time, in milliseconds since the epoch, when the job execution was last updated.

newJobExecution :: JobExecution

Create a value of JobExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startedAt:JobExecution', jobExecution_startedAt - The time, in milliseconds since the epoch, when the job execution was started.

$sel:status:JobExecution', jobExecution_status - The status of the job execution. Can be one of: "QUEUED", "IN_PROGRESS", "FAILED", "SUCCESS", "CANCELED", "REJECTED", or "REMOVED".

$sel:statusDetails:JobExecution', jobExecution_statusDetails - A collection of name/value pairs that describe the status of the job execution.

$sel:thingName:JobExecution', jobExecution_thingName - The name of the thing that is executing the job.

$sel:queuedAt:JobExecution', jobExecution_queuedAt - The time, in milliseconds since the epoch, when the job execution was enqueued.

$sel:executionNumber:JobExecution', jobExecution_executionNumber - A number that identifies a particular job execution on a particular device. It can be used later in commands that return or update job execution information.

$sel:versionNumber:JobExecution', jobExecution_versionNumber - The version of the job execution. Job execution versions are incremented each time they are updated by a device.

$sel:jobDocument:JobExecution', jobExecution_jobDocument - The content of the job document.

$sel:approximateSecondsBeforeTimedOut:JobExecution', jobExecution_approximateSecondsBeforeTimedOut - The estimated number of seconds that remain before the job execution status will be changed to TIMED_OUT.

$sel:jobId:JobExecution', jobExecution_jobId - The unique identifier you assigned to this job when it was created.

$sel:lastUpdatedAt:JobExecution', jobExecution_lastUpdatedAt - The time, in milliseconds since the epoch, when the job execution was last updated.

jobExecution_startedAt :: Lens' JobExecution (Maybe Integer)

The time, in milliseconds since the epoch, when the job execution was started.

jobExecution_status :: Lens' JobExecution (Maybe JobExecutionStatus)

The status of the job execution. Can be one of: "QUEUED", "IN_PROGRESS", "FAILED", "SUCCESS", "CANCELED", "REJECTED", or "REMOVED".

jobExecution_statusDetails :: Lens' JobExecution (Maybe (HashMap Text Text))

A collection of name/value pairs that describe the status of the job execution.

jobExecution_thingName :: Lens' JobExecution (Maybe Text)

The name of the thing that is executing the job.

jobExecution_queuedAt :: Lens' JobExecution (Maybe Integer)

The time, in milliseconds since the epoch, when the job execution was enqueued.

jobExecution_executionNumber :: Lens' JobExecution (Maybe Integer)

A number that identifies a particular job execution on a particular device. It can be used later in commands that return or update job execution information.

jobExecution_versionNumber :: Lens' JobExecution (Maybe Integer)

The version of the job execution. Job execution versions are incremented each time they are updated by a device.

jobExecution_jobDocument :: Lens' JobExecution (Maybe Text)

The content of the job document.

jobExecution_approximateSecondsBeforeTimedOut :: Lens' JobExecution (Maybe Integer)

The estimated number of seconds that remain before the job execution status will be changed to TIMED_OUT.

jobExecution_jobId :: Lens' JobExecution (Maybe Text)

The unique identifier you assigned to this job when it was created.

jobExecution_lastUpdatedAt :: Lens' JobExecution (Maybe Integer)

The time, in milliseconds since the epoch, when the job execution was last updated.

JobExecutionState

data JobExecutionState

Contains data about the state of a job execution.

See: newJobExecutionState smart constructor.

Constructors

JobExecutionState'

Fields

status :: Maybe JobExecutionStatus

The status of the job execution. Can be one of: "QUEUED", "IN_PROGRESS", "FAILED", "SUCCESS", "CANCELED", "REJECTED", or "REMOVED".

statusDetails :: Maybe (HashMap Text Text)

A collection of name/value pairs that describe the status of the job execution.

versionNumber :: Maybe Integer

The version of the job execution. Job execution versions are incremented each time they are updated by a device.

newJobExecutionState :: JobExecutionState

Create a value of JobExecutionState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:JobExecutionState', jobExecutionState_status - The status of the job execution. Can be one of: "QUEUED", "IN_PROGRESS", "FAILED", "SUCCESS", "CANCELED", "REJECTED", or "REMOVED".

$sel:statusDetails:JobExecutionState', jobExecutionState_statusDetails - A collection of name/value pairs that describe the status of the job execution.

$sel:versionNumber:JobExecutionState', jobExecutionState_versionNumber - The version of the job execution. Job execution versions are incremented each time they are updated by a device.

jobExecutionState_status :: Lens' JobExecutionState (Maybe JobExecutionStatus)

The status of the job execution. Can be one of: "QUEUED", "IN_PROGRESS", "FAILED", "SUCCESS", "CANCELED", "REJECTED", or "REMOVED".

jobExecutionState_statusDetails :: Lens' JobExecutionState (Maybe (HashMap Text Text))

A collection of name/value pairs that describe the status of the job execution.

jobExecutionState_versionNumber :: Lens' JobExecutionState (Maybe Integer)

The version of the job execution. Job execution versions are incremented each time they are updated by a device.

JobExecutionSummary

data JobExecutionSummary

Contains a subset of information about a job execution.

See: newJobExecutionSummary smart constructor.

Constructors

JobExecutionSummary'

Fields

startedAt :: Maybe Integer

The time, in milliseconds since the epoch, when the job execution started.

queuedAt :: Maybe Integer

The time, in milliseconds since the epoch, when the job execution was enqueued.

executionNumber :: Maybe Integer

A number that identifies a particular job execution on a particular device.

versionNumber :: Maybe Integer

The version of the job execution. Job execution versions are incremented each time AWS IoT Jobs receives an update from a device.

jobId :: Maybe Text

The unique identifier you assigned to this job when it was created.

lastUpdatedAt :: Maybe Integer

The time, in milliseconds since the epoch, when the job execution was last updated.

newJobExecutionSummary :: JobExecutionSummary

Create a value of JobExecutionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startedAt:JobExecutionSummary', jobExecutionSummary_startedAt - The time, in milliseconds since the epoch, when the job execution started.

$sel:queuedAt:JobExecutionSummary', jobExecutionSummary_queuedAt - The time, in milliseconds since the epoch, when the job execution was enqueued.

$sel:executionNumber:JobExecutionSummary', jobExecutionSummary_executionNumber - A number that identifies a particular job execution on a particular device.

$sel:versionNumber:JobExecutionSummary', jobExecutionSummary_versionNumber - The version of the job execution. Job execution versions are incremented each time AWS IoT Jobs receives an update from a device.

$sel:jobId:JobExecutionSummary', jobExecutionSummary_jobId - The unique identifier you assigned to this job when it was created.

$sel:lastUpdatedAt:JobExecutionSummary', jobExecutionSummary_lastUpdatedAt - The time, in milliseconds since the epoch, when the job execution was last updated.

jobExecutionSummary_startedAt :: Lens' JobExecutionSummary (Maybe Integer)

The time, in milliseconds since the epoch, when the job execution started.

jobExecutionSummary_queuedAt :: Lens' JobExecutionSummary (Maybe Integer)

The time, in milliseconds since the epoch, when the job execution was enqueued.

jobExecutionSummary_executionNumber :: Lens' JobExecutionSummary (Maybe Integer)

A number that identifies a particular job execution on a particular device.

jobExecutionSummary_versionNumber :: Lens' JobExecutionSummary (Maybe Integer)

The version of the job execution. Job execution versions are incremented each time AWS IoT Jobs receives an update from a device.

jobExecutionSummary_jobId :: Lens' JobExecutionSummary (Maybe Text)

The unique identifier you assigned to this job when it was created.

jobExecutionSummary_lastUpdatedAt :: Lens' JobExecutionSummary (Maybe Integer)

The time, in milliseconds since the epoch, when the job execution was last updated.