Network.AWS.IoTJobsData.UpdateJobExecution

Description

Updates the status of a job execution.

Creating a Request

data UpdateJobExecution

See: newUpdateJobExecution smart constructor.

Constructors

UpdateJobExecution'

Fields

includeJobDocument :: Maybe Bool

Optional. When set to true, the response contains the job document. The default is false.

expectedVersion :: Maybe Integer

Optional. The expected current version of the job execution. Each time you update the job execution, its version is incremented. If the version of the job execution stored in Jobs does not match, the update is rejected with a VersionMismatch error, and an ErrorResponse that contains the current job execution status data is returned. (This makes it unnecessary to perform a separate DescribeJobExecution request in order to obtain the job execution status data.)

includeJobExecutionState :: Maybe Bool

Optional. When included and set to true, the response contains the JobExecutionState data. The default is false.

statusDetails :: Maybe (HashMap Text Text)

Optional. A collection of name/value pairs that describe the status of the job execution. If not specified, the statusDetails are unchanged.

stepTimeoutInMinutes :: Maybe Integer

Specifies the amount of time this device has to finish execution of this job. If the job execution status is not set to a terminal state before this timer expires, or before the timer is reset (by again calling UpdateJobExecution, setting the status to IN_PROGRESS and specifying a new timeout value in this field) the job execution status will be automatically set to TIMED_OUT. Note that setting or resetting this timeout has no effect on that job execution timeout which may have been specified when the job was created (CreateJob using field timeoutConfig).

executionNumber :: Maybe Integer

Optional. A number that identifies a particular job execution on a particular device.

jobId :: Text

The unique identifier assigned to this job when it was created.

thingName :: Text

The name of the thing associated with the device.

status :: JobExecutionStatus

The new status for the job execution (IN_PROGRESS, FAILED, SUCCESS, or REJECTED). This must be specified on every update.

newUpdateJobExecution

Arguments

:: Text
$sel:jobId:UpdateJobExecution'

-> Text
$sel:thingName:UpdateJobExecution'

-> JobExecutionStatus
$sel:status:UpdateJobExecution'

-> UpdateJobExecution

Create a value of UpdateJobExecution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeJobDocument:UpdateJobExecution', updateJobExecution_includeJobDocument - Optional. When set to true, the response contains the job document. The default is false.

$sel:expectedVersion:UpdateJobExecution', updateJobExecution_expectedVersion - Optional. The expected current version of the job execution. Each time you update the job execution, its version is incremented. If the version of the job execution stored in Jobs does not match, the update is rejected with a VersionMismatch error, and an ErrorResponse that contains the current job execution status data is returned. (This makes it unnecessary to perform a separate DescribeJobExecution request in order to obtain the job execution status data.)

$sel:includeJobExecutionState:UpdateJobExecution', updateJobExecution_includeJobExecutionState - Optional. When included and set to true, the response contains the JobExecutionState data. The default is false.

$sel:statusDetails:UpdateJobExecution', updateJobExecution_statusDetails - Optional. A collection of name/value pairs that describe the status of the job execution. If not specified, the statusDetails are unchanged.

$sel:stepTimeoutInMinutes:UpdateJobExecution', updateJobExecution_stepTimeoutInMinutes - Specifies the amount of time this device has to finish execution of this job. If the job execution status is not set to a terminal state before this timer expires, or before the timer is reset (by again calling UpdateJobExecution, setting the status to IN_PROGRESS and specifying a new timeout value in this field) the job execution status will be automatically set to TIMED_OUT. Note that setting or resetting this timeout has no effect on that job execution timeout which may have been specified when the job was created (CreateJob using field timeoutConfig).

$sel:executionNumber:UpdateJobExecution', updateJobExecution_executionNumber - Optional. A number that identifies a particular job execution on a particular device.

$sel:jobId:UpdateJobExecution', updateJobExecution_jobId - The unique identifier assigned to this job when it was created.

$sel:thingName:UpdateJobExecution', updateJobExecution_thingName - The name of the thing associated with the device.

$sel:status:UpdateJobExecution', updateJobExecution_status - The new status for the job execution (IN_PROGRESS, FAILED, SUCCESS, or REJECTED). This must be specified on every update.

Request Lenses

updateJobExecution_includeJobDocument :: Lens' UpdateJobExecution (Maybe Bool)

Optional. When set to true, the response contains the job document. The default is false.

updateJobExecution_expectedVersion :: Lens' UpdateJobExecution (Maybe Integer)

Optional. The expected current version of the job execution. Each time you update the job execution, its version is incremented. If the version of the job execution stored in Jobs does not match, the update is rejected with a VersionMismatch error, and an ErrorResponse that contains the current job execution status data is returned. (This makes it unnecessary to perform a separate DescribeJobExecution request in order to obtain the job execution status data.)

updateJobExecution_includeJobExecutionState :: Lens' UpdateJobExecution (Maybe Bool)

Optional. When included and set to true, the response contains the JobExecutionState data. The default is false.

updateJobExecution_statusDetails :: Lens' UpdateJobExecution (Maybe (HashMap Text Text))

Optional. A collection of name/value pairs that describe the status of the job execution. If not specified, the statusDetails are unchanged.

updateJobExecution_stepTimeoutInMinutes :: Lens' UpdateJobExecution (Maybe Integer)

Specifies the amount of time this device has to finish execution of this job. If the job execution status is not set to a terminal state before this timer expires, or before the timer is reset (by again calling UpdateJobExecution, setting the status to IN_PROGRESS and specifying a new timeout value in this field) the job execution status will be automatically set to TIMED_OUT. Note that setting or resetting this timeout has no effect on that job execution timeout which may have been specified when the job was created (CreateJob using field timeoutConfig).

updateJobExecution_executionNumber :: Lens' UpdateJobExecution (Maybe Integer)

Optional. A number that identifies a particular job execution on a particular device.

updateJobExecution_jobId :: Lens' UpdateJobExecution Text

The unique identifier assigned to this job when it was created.

updateJobExecution_thingName :: Lens' UpdateJobExecution Text

The name of the thing associated with the device.

updateJobExecution_status :: Lens' UpdateJobExecution JobExecutionStatus

The new status for the job execution (IN_PROGRESS, FAILED, SUCCESS, or REJECTED). This must be specified on every update.

Destructuring the Response

data UpdateJobExecutionResponse

See: newUpdateJobExecutionResponse smart constructor.

Constructors

UpdateJobExecutionResponse'

Fields

executionState :: Maybe JobExecutionState

A JobExecutionState object.

jobDocument :: Maybe Text

The contents of the Job Documents.

httpStatus :: Int

The response's http status code.

newUpdateJobExecutionResponse

Arguments

:: Int
$sel:httpStatus:UpdateJobExecutionResponse'

-> UpdateJobExecutionResponse

Create a value of UpdateJobExecutionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:executionState:UpdateJobExecutionResponse', updateJobExecutionResponse_executionState - A JobExecutionState object.

$sel:jobDocument:UpdateJobExecutionResponse', updateJobExecutionResponse_jobDocument - The contents of the Job Documents.

$sel:httpStatus:UpdateJobExecutionResponse', updateJobExecutionResponse_httpStatus - The response's http status code.

Response Lenses

updateJobExecutionResponse_executionState :: Lens' UpdateJobExecutionResponse (Maybe JobExecutionState)

A JobExecutionState object.

updateJobExecutionResponse_jobDocument :: Lens' UpdateJobExecutionResponse (Maybe Text)

The contents of the Job Documents.

updateJobExecutionResponse_httpStatus :: Lens' UpdateJobExecutionResponse Int

The response's http status code.