Network.AWS.IoTSecureTunneling

Description

Derived from API version 2018-10-05 of the AWS service descriptions, licensed under Apache 2.0.

AWS IoT Secure Tunneling

AWS IoT Secure Tunnling enables you to create remote connections to devices deployed in the field.

For more information about how AWS IoT Secure Tunneling works, see AWS IoT Secure Tunneling.

Service Configuration

defaultService :: Service

API version 2018-10-05 of the Amazon IoT Secure Tunneling SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by IoTSecureTunneling.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown when an operation is attempted on a resource that does not exist.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Thrown when a tunnel limit is exceeded.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resourceArn:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ListTagsForResource', listTagsForResource_resourceArn - The resource ARN.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (NonEmpty Tag)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - The tags for the specified resource.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

CloseTunnel

data CloseTunnel

See: newCloseTunnel smart constructor.

Constructors

CloseTunnel' (Maybe Bool) Text

newCloseTunnel

Arguments

:: Text
$sel:tunnelId:CloseTunnel'

-> CloseTunnel

Create a value of CloseTunnel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:delete':CloseTunnel', closeTunnel_delete - When set to true, AWS IoT Secure Tunneling deletes the tunnel data immediately.

$sel:tunnelId:CloseTunnel', closeTunnel_tunnelId - The ID of the tunnel to close.

data CloseTunnelResponse

See: newCloseTunnelResponse smart constructor.

Constructors

CloseTunnelResponse' Int

newCloseTunnelResponse

Arguments

:: Int
$sel:httpStatus:CloseTunnelResponse'

-> CloseTunnelResponse

Create a value of CloseTunnelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CloseTunnelResponse', closeTunnelResponse_httpStatus - The response's http status code.

OpenTunnel

data OpenTunnel

See: newOpenTunnel smart constructor.

Constructors

OpenTunnel' (Maybe DestinationConfig) (Maybe Text) (Maybe TimeoutConfig) (Maybe (NonEmpty Tag))

newOpenTunnel :: OpenTunnel

Create a value of OpenTunnel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationConfig:OpenTunnel', openTunnel_destinationConfig - The destination configuration for the OpenTunnel request.

$sel:description:OpenTunnel', openTunnel_description - A short text description of the tunnel.

$sel:timeoutConfig:OpenTunnel', openTunnel_timeoutConfig - Timeout configuration for a tunnel.

$sel:tags:OpenTunnel', openTunnel_tags - A collection of tag metadata.

data OpenTunnelResponse

See: newOpenTunnelResponse smart constructor.

Constructors

OpenTunnelResponse' (Maybe (Sensitive Text)) (Maybe Text) (Maybe (Sensitive Text)) (Maybe Text) Int

newOpenTunnelResponse

Arguments

:: Int
$sel:httpStatus:OpenTunnelResponse'

-> OpenTunnelResponse

Create a value of OpenTunnelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceAccessToken:OpenTunnelResponse', openTunnelResponse_sourceAccessToken - The access token the source local proxy uses to connect to AWS IoT Secure Tunneling.

$sel:tunnelArn:OpenTunnelResponse', openTunnelResponse_tunnelArn - The Amazon Resource Name for the tunnel. The tunnel ARN format is arn:aws:tunnel:<region>:<account-id>:tunnel/<tunnel-id>

$sel:destinationAccessToken:OpenTunnelResponse', openTunnelResponse_destinationAccessToken - The access token the destination local proxy uses to connect to AWS IoT Secure Tunneling.

$sel:tunnelId:OpenTunnelResponse', openTunnelResponse_tunnelId - A unique alpha-numeric tunnel ID.

$sel:httpStatus:OpenTunnelResponse', openTunnelResponse_httpStatus - The response's http status code.

DescribeTunnel

data DescribeTunnel

See: newDescribeTunnel smart constructor.

Constructors

DescribeTunnel' Text

newDescribeTunnel

Arguments

:: Text
$sel:tunnelId:DescribeTunnel'

-> DescribeTunnel

Create a value of DescribeTunnel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tunnelId:DescribeTunnel', describeTunnel_tunnelId - The tunnel to describe.

data DescribeTunnelResponse

See: newDescribeTunnelResponse smart constructor.

Constructors

DescribeTunnelResponse' (Maybe Tunnel) Int

newDescribeTunnelResponse

Arguments

:: Int
$sel:httpStatus:DescribeTunnelResponse'

-> DescribeTunnelResponse

Create a value of DescribeTunnelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tunnel:DescribeTunnelResponse', describeTunnelResponse_tunnel - The tunnel being described.

$sel:httpStatus:DescribeTunnelResponse', describeTunnelResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (NonEmpty Tag)

newTagResource

Arguments

:: Text
$sel:resourceArn:TagResource'

-> NonEmpty Tag
$sel:tags:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:TagResource', tagResource_resourceArn - The ARN of the resource.

$sel:tags:TagResource', tagResource_tags - The tags for the resource.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text [Text]

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:UntagResource', untagResource_resourceArn - The resource ARN.

$sel:tagKeys:UntagResource', untagResource_tagKeys - The keys of the tags to remove.

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

ListTunnels

data ListTunnels

See: newListTunnels smart constructor.

Constructors

ListTunnels' (Maybe Text) (Maybe Text) (Maybe Natural)

newListTunnels :: ListTunnels

Create a value of ListTunnels with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTunnels', listTunnels_nextToken - A token to retrieve the next set of results.

$sel:thingName:ListTunnels', listTunnels_thingName - The name of the IoT thing associated with the destination device.

$sel:maxResults:ListTunnels', listTunnels_maxResults - The maximum number of results to return at once.

data ListTunnelsResponse

See: newListTunnelsResponse smart constructor.

Constructors

ListTunnelsResponse' (Maybe Text) (Maybe [TunnelSummary]) Int

newListTunnelsResponse

Arguments

:: Int
$sel:httpStatus:ListTunnelsResponse'

-> ListTunnelsResponse

Create a value of ListTunnelsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTunnels', listTunnelsResponse_nextToken - A token to used to retrieve the next set of results.

$sel:tunnelSummaries:ListTunnelsResponse', listTunnelsResponse_tunnelSummaries - A short description of the tunnels in an AWS account.

$sel:httpStatus:ListTunnelsResponse', listTunnelsResponse_httpStatus - The response's http status code.

Types

ConnectionStatus

newtype ConnectionStatus

Constructors

ConnectionStatus'

Fields

fromConnectionStatus :: Text

Bundled Patterns

pattern ConnectionStatus_CONNECTED :: ConnectionStatus

pattern ConnectionStatus_DISCONNECTED :: ConnectionStatus

TunnelStatus

newtype TunnelStatus

Constructors

TunnelStatus'

Fields

fromTunnelStatus :: Text

Bundled Patterns

pattern TunnelStatus_CLOSED :: TunnelStatus

pattern TunnelStatus_OPEN :: TunnelStatus

ConnectionState

data ConnectionState

The state of a connection.

See: newConnectionState smart constructor.

Constructors

ConnectionState' (Maybe ConnectionStatus) (Maybe POSIX)

newConnectionState :: ConnectionState

Create a value of ConnectionState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ConnectionState', connectionState_status - The connection status of the tunnel. Valid values are CONNECTED and DISCONNECTED.

$sel:lastUpdatedAt:ConnectionState', connectionState_lastUpdatedAt - The last time the connection status was updated.

DestinationConfig

data DestinationConfig

The destination configuration.

See: newDestinationConfig smart constructor.

Constructors

DestinationConfig' (Maybe Text) (NonEmpty Text)

newDestinationConfig

Arguments

:: NonEmpty Text
$sel:services:DestinationConfig'

-> DestinationConfig

Create a value of DestinationConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thingName:DestinationConfig', destinationConfig_thingName - The name of the IoT thing to which you want to connect.

$sel:services:DestinationConfig', destinationConfig_services - A list of service names that identity the target application. The AWS IoT client running on the destination device reads this value and uses it to look up a port or an IP address and a port. The AWS IoT client instantiates the local proxy which uses this information to connect to the destination application.

Tag

data Tag

An arbitary key/value pair used to add searchable metadata to secure tunnel resources.

See: newTag smart constructor.

Constructors

Tag' Text Text

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The key of the tag.

$sel:value:Tag', tag_value - The value of the tag.

TimeoutConfig

data TimeoutConfig

Tunnel timeout configuration.

See: newTimeoutConfig smart constructor.

Constructors

TimeoutConfig' (Maybe Natural)

newTimeoutConfig :: TimeoutConfig

Create a value of TimeoutConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxLifetimeTimeoutMinutes:TimeoutConfig', timeoutConfig_maxLifetimeTimeoutMinutes - The maximum amount of time (in minutes) a tunnel can remain open. If not specified, maxLifetimeTimeoutMinutes defaults to 720 minutes. Valid values are from 1 minute to 12 hours (720 minutes)

Tunnel

data Tunnel

A connection between a source computer and a destination device.

See: newTunnel smart constructor.

Constructors

Tunnel' (Maybe TunnelStatus) (Maybe POSIX) (Maybe POSIX) (Maybe Text) (Maybe ConnectionState) (Maybe ConnectionState) (Maybe Text) (Maybe DestinationConfig) (Maybe Text) (Maybe TimeoutConfig) (Maybe (NonEmpty Tag))

newTunnel :: Tunnel

Create a value of Tunnel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Tunnel', tunnel_status - The status of a tunnel. Valid values are: Open and Closed.

$sel:lastUpdatedAt:Tunnel', tunnel_lastUpdatedAt - The last time the tunnel was updated.

$sel:createdAt:Tunnel', tunnel_createdAt - The time when the tunnel was created.

$sel:tunnelArn:Tunnel', tunnel_tunnelArn - The Amazon Resource Name (ARN) of a tunnel. The tunnel ARN format is arn:aws:tunnel:<region>:<account-id>:tunnel/<tunnel-id>

$sel:sourceConnectionState:Tunnel', tunnel_sourceConnectionState - The connection state of the source application.

$sel:destinationConnectionState:Tunnel', tunnel_destinationConnectionState - The connection state of the destination application.

$sel:tunnelId:Tunnel', tunnel_tunnelId - A unique alpha-numeric ID that identifies a tunnel.

$sel:destinationConfig:Tunnel', tunnel_destinationConfig - The destination configuration that specifies the thing name of the destination device and a service name that the local proxy uses to connect to the destination application.

$sel:description:Tunnel', tunnel_description - A description of the tunnel.

$sel:timeoutConfig:Tunnel', tunnel_timeoutConfig - Timeout configuration for the tunnel.

$sel:tags:Tunnel', tunnel_tags - A list of tag metadata associated with the secure tunnel.

TunnelSummary

data TunnelSummary

Information about the tunnel.

See: newTunnelSummary smart constructor.

Constructors

TunnelSummary' (Maybe TunnelStatus) (Maybe POSIX) (Maybe POSIX) (Maybe Text) (Maybe Text) (Maybe Text)

newTunnelSummary :: TunnelSummary

Create a value of TunnelSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:TunnelSummary', tunnelSummary_status - The status of a tunnel. Valid values are: Open and Closed.

$sel:lastUpdatedAt:TunnelSummary', tunnelSummary_lastUpdatedAt - The time the tunnel was last updated.

$sel:createdAt:TunnelSummary', tunnelSummary_createdAt - The time the tunnel was created.

$sel:tunnelArn:TunnelSummary', tunnelSummary_tunnelArn - The Amazon Resource Name of the tunnel. The tunnel ARN format is arn:aws:tunnel:<region>:<account-id>:tunnel/<tunnel-id>

$sel:tunnelId:TunnelSummary', tunnelSummary_tunnelId - The unique alpha-numeric identifier for the tunnel.

$sel:description:TunnelSummary', tunnelSummary_description - A description of the tunnel.