Network.AWS.IoTSecureTunneling.OpenTunnel

Description

Creates a new tunnel, and returns two client access tokens for clients to use to connect to the AWS IoT Secure Tunneling proxy server.

Creating a Request

data OpenTunnel

See: newOpenTunnel smart constructor.

Constructors

OpenTunnel'

Fields

destinationConfig :: Maybe DestinationConfig

The destination configuration for the OpenTunnel request.

description :: Maybe Text

A short text description of the tunnel.

timeoutConfig :: Maybe TimeoutConfig

Timeout configuration for a tunnel.

tags :: Maybe (NonEmpty Tag)

A collection of tag metadata.

newOpenTunnel :: OpenTunnel

Create a value of OpenTunnel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationConfig:OpenTunnel', openTunnel_destinationConfig - The destination configuration for the OpenTunnel request.

$sel:description:OpenTunnel', openTunnel_description - A short text description of the tunnel.

$sel:timeoutConfig:OpenTunnel', openTunnel_timeoutConfig - Timeout configuration for a tunnel.

$sel:tags:OpenTunnel', openTunnel_tags - A collection of tag metadata.

Request Lenses

openTunnel_destinationConfig :: Lens' OpenTunnel (Maybe DestinationConfig)

The destination configuration for the OpenTunnel request.

openTunnel_description :: Lens' OpenTunnel (Maybe Text)

A short text description of the tunnel.

openTunnel_timeoutConfig :: Lens' OpenTunnel (Maybe TimeoutConfig)

Timeout configuration for a tunnel.

openTunnel_tags :: Lens' OpenTunnel (Maybe (NonEmpty Tag))

A collection of tag metadata.

Destructuring the Response

data OpenTunnelResponse

See: newOpenTunnelResponse smart constructor.

Constructors

OpenTunnelResponse'

Fields

sourceAccessToken :: Maybe (Sensitive Text)

The access token the source local proxy uses to connect to AWS IoT Secure Tunneling.

tunnelArn :: Maybe Text

The Amazon Resource Name for the tunnel. The tunnel ARN format is arn:aws:tunnel:<region>:<account-id>:tunnel/<tunnel-id>

destinationAccessToken :: Maybe (Sensitive Text)

The access token the destination local proxy uses to connect to AWS IoT Secure Tunneling.

tunnelId :: Maybe Text

A unique alpha-numeric tunnel ID.

httpStatus :: Int

The response's http status code.

newOpenTunnelResponse

Arguments

:: Int
$sel:httpStatus:OpenTunnelResponse'

-> OpenTunnelResponse

Create a value of OpenTunnelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceAccessToken:OpenTunnelResponse', openTunnelResponse_sourceAccessToken - The access token the source local proxy uses to connect to AWS IoT Secure Tunneling.

$sel:tunnelArn:OpenTunnelResponse', openTunnelResponse_tunnelArn - The Amazon Resource Name for the tunnel. The tunnel ARN format is arn:aws:tunnel:<region>:<account-id>:tunnel/<tunnel-id>

$sel:destinationAccessToken:OpenTunnelResponse', openTunnelResponse_destinationAccessToken - The access token the destination local proxy uses to connect to AWS IoT Secure Tunneling.

$sel:tunnelId:OpenTunnelResponse', openTunnelResponse_tunnelId - A unique alpha-numeric tunnel ID.

$sel:httpStatus:OpenTunnelResponse', openTunnelResponse_httpStatus - The response's http status code.

Response Lenses

openTunnelResponse_sourceAccessToken :: Lens' OpenTunnelResponse (Maybe Text)

The access token the source local proxy uses to connect to AWS IoT Secure Tunneling.

openTunnelResponse_tunnelArn :: Lens' OpenTunnelResponse (Maybe Text)

The Amazon Resource Name for the tunnel. The tunnel ARN format is arn:aws:tunnel:<region>:<account-id>:tunnel/<tunnel-id>

openTunnelResponse_destinationAccessToken :: Lens' OpenTunnelResponse (Maybe Text)

The access token the destination local proxy uses to connect to AWS IoT Secure Tunneling.

openTunnelResponse_tunnelId :: Lens' OpenTunnelResponse (Maybe Text)

A unique alpha-numeric tunnel ID.

openTunnelResponse_httpStatus :: Lens' OpenTunnelResponse Int

The response's http status code.