Network.AWS.IoTSecureTunneling.Types.Tunnel

Description

Documentation

data Tunnel

A connection between a source computer and a destination device.

See: newTunnel smart constructor.

Constructors

Tunnel'

Fields

status :: Maybe TunnelStatus

The status of a tunnel. Valid values are: Open and Closed.

lastUpdatedAt :: Maybe POSIX

The last time the tunnel was updated.

createdAt :: Maybe POSIX

The time when the tunnel was created.

tunnelArn :: Maybe Text

The Amazon Resource Name (ARN) of a tunnel. The tunnel ARN format is arn:aws:tunnel:<region>:<account-id>:tunnel/<tunnel-id>

sourceConnectionState :: Maybe ConnectionState

The connection state of the source application.

destinationConnectionState :: Maybe ConnectionState

The connection state of the destination application.

tunnelId :: Maybe Text

A unique alpha-numeric ID that identifies a tunnel.

destinationConfig :: Maybe DestinationConfig

The destination configuration that specifies the thing name of the destination device and a service name that the local proxy uses to connect to the destination application.

description :: Maybe Text

A description of the tunnel.

timeoutConfig :: Maybe TimeoutConfig

Timeout configuration for the tunnel.

tags :: Maybe (NonEmpty Tag)

A list of tag metadata associated with the secure tunnel.

newTunnel :: Tunnel

Create a value of Tunnel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Tunnel', tunnel_status - The status of a tunnel. Valid values are: Open and Closed.

$sel:lastUpdatedAt:Tunnel', tunnel_lastUpdatedAt - The last time the tunnel was updated.

$sel:createdAt:Tunnel', tunnel_createdAt - The time when the tunnel was created.

$sel:tunnelArn:Tunnel', tunnel_tunnelArn - The Amazon Resource Name (ARN) of a tunnel. The tunnel ARN format is arn:aws:tunnel:<region>:<account-id>:tunnel/<tunnel-id>

$sel:sourceConnectionState:Tunnel', tunnel_sourceConnectionState - The connection state of the source application.

$sel:destinationConnectionState:Tunnel', tunnel_destinationConnectionState - The connection state of the destination application.

$sel:tunnelId:Tunnel', tunnel_tunnelId - A unique alpha-numeric ID that identifies a tunnel.

$sel:destinationConfig:Tunnel', tunnel_destinationConfig - The destination configuration that specifies the thing name of the destination device and a service name that the local proxy uses to connect to the destination application.

$sel:description:Tunnel', tunnel_description - A description of the tunnel.

$sel:timeoutConfig:Tunnel', tunnel_timeoutConfig - Timeout configuration for the tunnel.

$sel:tags:Tunnel', tunnel_tags - A list of tag metadata associated with the secure tunnel.

tunnel_status :: Lens' Tunnel (Maybe TunnelStatus)

The status of a tunnel. Valid values are: Open and Closed.

tunnel_lastUpdatedAt :: Lens' Tunnel (Maybe UTCTime)

The last time the tunnel was updated.

tunnel_createdAt :: Lens' Tunnel (Maybe UTCTime)

The time when the tunnel was created.

tunnel_tunnelArn :: Lens' Tunnel (Maybe Text)

The Amazon Resource Name (ARN) of a tunnel. The tunnel ARN format is arn:aws:tunnel:<region>:<account-id>:tunnel/<tunnel-id>

tunnel_sourceConnectionState :: Lens' Tunnel (Maybe ConnectionState)

The connection state of the source application.

tunnel_destinationConnectionState :: Lens' Tunnel (Maybe ConnectionState)

The connection state of the destination application.

tunnel_tunnelId :: Lens' Tunnel (Maybe Text)

A unique alpha-numeric ID that identifies a tunnel.

tunnel_destinationConfig :: Lens' Tunnel (Maybe DestinationConfig)

The destination configuration that specifies the thing name of the destination device and a service name that the local proxy uses to connect to the destination application.

tunnel_description :: Lens' Tunnel (Maybe Text)

A description of the tunnel.

tunnel_timeoutConfig :: Lens' Tunnel (Maybe TimeoutConfig)

Timeout configuration for the tunnel.

tunnel_tags :: Lens' Tunnel (Maybe (NonEmpty Tag))

A list of tag metadata associated with the secure tunnel.