Network.AWS.IoTSiteWise.CreateAccessPolicy

Description

Creates an access policy that grants the specified identity (Amazon Web Services SSO user, Amazon Web Services SSO group, or IAM user) access to the specified IoT SiteWise Monitor portal or project resource.

Creating a Request

data CreateAccessPolicy

See: newCreateAccessPolicy smart constructor.

Constructors

CreateAccessPolicy'

Fields

clientToken :: Maybe Text

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

tags :: Maybe (HashMap Text Text)

A list of key-value pairs that contain metadata for the access policy. For more information, see Tagging your IoT SiteWise resources in the IoT SiteWise User Guide.

accessPolicyIdentity :: Identity

The identity for this access policy. Choose an Amazon Web Services SSO user, an Amazon Web Services SSO group, or an IAM user.

accessPolicyResource :: Resource

The IoT SiteWise Monitor resource for this access policy. Choose either a portal or a project.

accessPolicyPermission :: Permission

The permission level for this access policy. Note that a project ADMINISTRATOR is also known as a project owner.

newCreateAccessPolicy

Arguments

:: Identity
$sel:accessPolicyIdentity:CreateAccessPolicy'

-> Resource
$sel:accessPolicyResource:CreateAccessPolicy'

-> Permission
$sel:accessPolicyPermission:CreateAccessPolicy'

-> CreateAccessPolicy

Create a value of CreateAccessPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateAccessPolicy', createAccessPolicy_clientToken - A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

$sel:tags:CreateAccessPolicy', createAccessPolicy_tags - A list of key-value pairs that contain metadata for the access policy. For more information, see Tagging your IoT SiteWise resources in the IoT SiteWise User Guide.

$sel:accessPolicyIdentity:CreateAccessPolicy', createAccessPolicy_accessPolicyIdentity - The identity for this access policy. Choose an Amazon Web Services SSO user, an Amazon Web Services SSO group, or an IAM user.

$sel:accessPolicyResource:CreateAccessPolicy', createAccessPolicy_accessPolicyResource - The IoT SiteWise Monitor resource for this access policy. Choose either a portal or a project.

$sel:accessPolicyPermission:CreateAccessPolicy', createAccessPolicy_accessPolicyPermission - The permission level for this access policy. Note that a project ADMINISTRATOR is also known as a project owner.

Request Lenses

createAccessPolicy_clientToken :: Lens' CreateAccessPolicy (Maybe Text)

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

createAccessPolicy_tags :: Lens' CreateAccessPolicy (Maybe (HashMap Text Text))

A list of key-value pairs that contain metadata for the access policy. For more information, see Tagging your IoT SiteWise resources in the IoT SiteWise User Guide.

createAccessPolicy_accessPolicyIdentity :: Lens' CreateAccessPolicy Identity

The identity for this access policy. Choose an Amazon Web Services SSO user, an Amazon Web Services SSO group, or an IAM user.

createAccessPolicy_accessPolicyResource :: Lens' CreateAccessPolicy Resource

The IoT SiteWise Monitor resource for this access policy. Choose either a portal or a project.

createAccessPolicy_accessPolicyPermission :: Lens' CreateAccessPolicy Permission

The permission level for this access policy. Note that a project ADMINISTRATOR is also known as a project owner.

Destructuring the Response

data CreateAccessPolicyResponse

See: newCreateAccessPolicyResponse smart constructor.

Constructors

CreateAccessPolicyResponse'

Fields

httpStatus :: Int

The response's http status code.

accessPolicyId :: Text

The ID of the access policy.

accessPolicyArn :: Text

The ARN of the access policy, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:access-policy/${AccessPolicyId}

newCreateAccessPolicyResponse

Arguments

:: Int
$sel:httpStatus:CreateAccessPolicyResponse'

-> Text
$sel:accessPolicyId:CreateAccessPolicyResponse'

-> Text
$sel:accessPolicyArn:CreateAccessPolicyResponse'

-> CreateAccessPolicyResponse

Create a value of CreateAccessPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateAccessPolicyResponse', createAccessPolicyResponse_httpStatus - The response's http status code.

$sel:accessPolicyId:CreateAccessPolicyResponse', createAccessPolicyResponse_accessPolicyId - The ID of the access policy.

$sel:accessPolicyArn:CreateAccessPolicyResponse', createAccessPolicyResponse_accessPolicyArn - The ARN of the access policy, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:access-policy/${AccessPolicyId}

Response Lenses

createAccessPolicyResponse_httpStatus :: Lens' CreateAccessPolicyResponse Int

The response's http status code.

createAccessPolicyResponse_accessPolicyId :: Lens' CreateAccessPolicyResponse Text

The ID of the access policy.

createAccessPolicyResponse_accessPolicyArn :: Lens' CreateAccessPolicyResponse Text

The ARN of the access policy, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:access-policy/${AccessPolicyId}