Network.AWS.IoTSiteWise.CreateGateway

Description

Creates a gateway, which is a virtual or edge device that delivers industrial data streams from local servers to IoT SiteWise. For more information, see Ingesting data using a gateway in the IoT SiteWise User Guide.

Creating a Request

data CreateGateway

See: newCreateGateway smart constructor.

Constructors

CreateGateway'

Fields

tags :: Maybe (HashMap Text Text)

A list of key-value pairs that contain metadata for the gateway. For more information, see Tagging your IoT SiteWise resources in the IoT SiteWise User Guide.

gatewayName :: Text

A unique, friendly name for the gateway.

gatewayPlatform :: GatewayPlatform

The gateway's platform. You can only specify one platform in a gateway.

newCreateGateway

Arguments

:: Text
$sel:gatewayName:CreateGateway'

-> GatewayPlatform
$sel:gatewayPlatform:CreateGateway'

-> CreateGateway

Create a value of CreateGateway with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateGateway', createGateway_tags - A list of key-value pairs that contain metadata for the gateway. For more information, see Tagging your IoT SiteWise resources in the IoT SiteWise User Guide.

$sel:gatewayName:CreateGateway', createGateway_gatewayName - A unique, friendly name for the gateway.

$sel:gatewayPlatform:CreateGateway', createGateway_gatewayPlatform - The gateway's platform. You can only specify one platform in a gateway.

Request Lenses

createGateway_tags :: Lens' CreateGateway (Maybe (HashMap Text Text))

A list of key-value pairs that contain metadata for the gateway. For more information, see Tagging your IoT SiteWise resources in the IoT SiteWise User Guide.

createGateway_gatewayName :: Lens' CreateGateway Text

A unique, friendly name for the gateway.

createGateway_gatewayPlatform :: Lens' CreateGateway GatewayPlatform

The gateway's platform. You can only specify one platform in a gateway.

Destructuring the Response

data CreateGatewayResponse

See: newCreateGatewayResponse smart constructor.

Constructors

CreateGatewayResponse'

Fields

httpStatus :: Int

The response's http status code.

gatewayId :: Text

The ID of the gateway device. You can use this ID when you call other IoT SiteWise APIs.

gatewayArn :: Text

The ARN of the gateway, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:gateway/${GatewayId}

newCreateGatewayResponse

Arguments

:: Int
$sel:httpStatus:CreateGatewayResponse'

-> Text
$sel:gatewayId:CreateGatewayResponse'

-> Text
$sel:gatewayArn:CreateGatewayResponse'

-> CreateGatewayResponse

Create a value of CreateGatewayResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateGatewayResponse', createGatewayResponse_httpStatus - The response's http status code.

$sel:gatewayId:CreateGatewayResponse', createGatewayResponse_gatewayId - The ID of the gateway device. You can use this ID when you call other IoT SiteWise APIs.

$sel:gatewayArn:CreateGatewayResponse', createGatewayResponse_gatewayArn - The ARN of the gateway, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:gateway/${GatewayId}

Response Lenses

createGatewayResponse_httpStatus :: Lens' CreateGatewayResponse Int

The response's http status code.

createGatewayResponse_gatewayId :: Lens' CreateGatewayResponse Text

The ID of the gateway device. You can use this ID when you call other IoT SiteWise APIs.

createGatewayResponse_gatewayArn :: Lens' CreateGatewayResponse Text

The ARN of the gateway, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:gateway/${GatewayId}