Network.AWS.IoTSiteWise.CreatePortal

Description

Creates a portal, which can contain projects and dashboards. IoT SiteWise Monitor uses Amazon Web Services SSO or IAM to authenticate portal users and manage user permissions.

Before you can sign in to a new portal, you must add at least one identity to that portal. For more information, see Adding or removing portal administrators in the IoT SiteWise User Guide.

Creating a Request

data CreatePortal

See: newCreatePortal smart constructor.

Constructors

CreatePortal'

Fields

portalAuthMode :: Maybe AuthMode

The service to use to authenticate users to the portal. Choose from the following options:

SSO – The portal uses Amazon Web Services Single Sign On to authenticate users and manage user permissions. Before you can create a portal that uses Amazon Web Services SSO, you must enable Amazon Web Services SSO. For more information, see Enabling Amazon Web Services SSO in the IoT SiteWise User Guide. This option is only available in Amazon Web Services Regions other than the China Regions.

IAM – The portal uses Identity and Access Management to authenticate users and manage user permissions. This option is only available in the China Regions.

You can't change this value after you create a portal.

Default: SSO

clientToken :: Maybe Text

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

portalDescription :: Maybe Text

A description for the portal.

notificationSenderEmail :: Maybe Text

The email address that sends alarm notifications.

If you use the IoT Events managed Lambda function to manage your emails, you must verify the sender email address in Amazon SES.

alarms :: Maybe Alarms

Contains the configuration information of an alarm created in an IoT SiteWise Monitor portal. You can use the alarm to monitor an asset property and get notified when the asset property value is outside a specified range. For more information, see Monitoring with alarms in the IoT SiteWise Application Guide.

portalLogoImageFile :: Maybe ImageFile

A logo image to display in the portal. Upload a square, high-resolution image. The image is displayed on a dark background.

tags :: Maybe (HashMap Text Text)

A list of key-value pairs that contain metadata for the portal. For more information, see Tagging your IoT SiteWise resources in the IoT SiteWise User Guide.

portalName :: Text

A friendly name for the portal.

portalContactEmail :: Text

The Amazon Web Services administrator's contact email address.

roleArn :: Text

The ARN of a service role that allows the portal's users to access your IoT SiteWise resources on your behalf. For more information, see Using service roles for IoT SiteWise Monitor in the IoT SiteWise User Guide.

newCreatePortal

Arguments

:: Text
$sel:portalName:CreatePortal'

-> Text
$sel:portalContactEmail:CreatePortal'

-> Text
$sel:roleArn:CreatePortal'

-> CreatePortal

Create a value of CreatePortal with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:portalAuthMode:CreatePortal', createPortal_portalAuthMode - The service to use to authenticate users to the portal. Choose from the following options:

SSO – The portal uses Amazon Web Services Single Sign On to authenticate users and manage user permissions. Before you can create a portal that uses Amazon Web Services SSO, you must enable Amazon Web Services SSO. For more information, see Enabling Amazon Web Services SSO in the IoT SiteWise User Guide. This option is only available in Amazon Web Services Regions other than the China Regions.

IAM – The portal uses Identity and Access Management to authenticate users and manage user permissions. This option is only available in the China Regions.

You can't change this value after you create a portal.

Default: SSO

$sel:clientToken:CreatePortal', createPortal_clientToken - A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

$sel:portalDescription:CreatePortal', createPortal_portalDescription - A description for the portal.

$sel:notificationSenderEmail:CreatePortal', createPortal_notificationSenderEmail - The email address that sends alarm notifications.

If you use the IoT Events managed Lambda function to manage your emails, you must verify the sender email address in Amazon SES.

$sel:alarms:CreatePortal', createPortal_alarms - Contains the configuration information of an alarm created in an IoT SiteWise Monitor portal. You can use the alarm to monitor an asset property and get notified when the asset property value is outside a specified range. For more information, see Monitoring with alarms in the IoT SiteWise Application Guide.

$sel:portalLogoImageFile:CreatePortal', createPortal_portalLogoImageFile - A logo image to display in the portal. Upload a square, high-resolution image. The image is displayed on a dark background.

$sel:tags:CreatePortal', createPortal_tags - A list of key-value pairs that contain metadata for the portal. For more information, see Tagging your IoT SiteWise resources in the IoT SiteWise User Guide.

$sel:portalName:CreatePortal', createPortal_portalName - A friendly name for the portal.

$sel:portalContactEmail:CreatePortal', createPortal_portalContactEmail - The Amazon Web Services administrator's contact email address.

$sel:roleArn:CreatePortal', createPortal_roleArn - The ARN of a service role that allows the portal's users to access your IoT SiteWise resources on your behalf. For more information, see Using service roles for IoT SiteWise Monitor in the IoT SiteWise User Guide.

Request Lenses

createPortal_portalAuthMode :: Lens' CreatePortal (Maybe AuthMode)

The service to use to authenticate users to the portal. Choose from the following options:

SSO – The portal uses Amazon Web Services Single Sign On to authenticate users and manage user permissions. Before you can create a portal that uses Amazon Web Services SSO, you must enable Amazon Web Services SSO. For more information, see Enabling Amazon Web Services SSO in the IoT SiteWise User Guide. This option is only available in Amazon Web Services Regions other than the China Regions.

IAM – The portal uses Identity and Access Management to authenticate users and manage user permissions. This option is only available in the China Regions.

You can't change this value after you create a portal.

Default: SSO

createPortal_clientToken :: Lens' CreatePortal (Maybe Text)

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

createPortal_portalDescription :: Lens' CreatePortal (Maybe Text)

A description for the portal.

createPortal_notificationSenderEmail :: Lens' CreatePortal (Maybe Text)

The email address that sends alarm notifications.

If you use the IoT Events managed Lambda function to manage your emails, you must verify the sender email address in Amazon SES.

createPortal_alarms :: Lens' CreatePortal (Maybe Alarms)

Contains the configuration information of an alarm created in an IoT SiteWise Monitor portal. You can use the alarm to monitor an asset property and get notified when the asset property value is outside a specified range. For more information, see Monitoring with alarms in the IoT SiteWise Application Guide.

createPortal_portalLogoImageFile :: Lens' CreatePortal (Maybe ImageFile)

A logo image to display in the portal. Upload a square, high-resolution image. The image is displayed on a dark background.

createPortal_tags :: Lens' CreatePortal (Maybe (HashMap Text Text))

A list of key-value pairs that contain metadata for the portal. For more information, see Tagging your IoT SiteWise resources in the IoT SiteWise User Guide.

createPortal_portalName :: Lens' CreatePortal Text

A friendly name for the portal.

createPortal_portalContactEmail :: Lens' CreatePortal Text

The Amazon Web Services administrator's contact email address.

createPortal_roleArn :: Lens' CreatePortal Text

The ARN of a service role that allows the portal's users to access your IoT SiteWise resources on your behalf. For more information, see Using service roles for IoT SiteWise Monitor in the IoT SiteWise User Guide.

Destructuring the Response

data CreatePortalResponse

See: newCreatePortalResponse smart constructor.

Constructors

CreatePortalResponse'

Fields

httpStatus :: Int

The response's http status code.

portalId :: Text

The ID of the created portal.

portalArn :: Text

The ARN of the portal, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:portal/${PortalId}

portalStartUrl :: Text

The URL for the IoT SiteWise Monitor portal. You can use this URL to access portals that use Amazon Web Services SSO for authentication. For portals that use IAM for authentication, you must use the IoT SiteWise console to get a URL that you can use to access the portal.

portalStatus :: PortalStatus

The status of the portal, which contains a state (CREATING after successfully calling this operation) and any error message.

ssoApplicationId :: Text

The associated Amazon Web Services SSO application ID, if the portal uses Amazon Web Services SSO.

newCreatePortalResponse

Arguments

:: Int
$sel:httpStatus:CreatePortalResponse'

-> Text
$sel:portalId:CreatePortalResponse'

-> Text
$sel:portalArn:CreatePortalResponse'

-> Text
$sel:portalStartUrl:CreatePortalResponse'

-> PortalStatus
$sel:portalStatus:CreatePortalResponse'

-> Text
$sel:ssoApplicationId:CreatePortalResponse'

-> CreatePortalResponse

Create a value of CreatePortalResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreatePortalResponse', createPortalResponse_httpStatus - The response's http status code.

$sel:portalId:CreatePortalResponse', createPortalResponse_portalId - The ID of the created portal.

$sel:portalArn:CreatePortalResponse', createPortalResponse_portalArn - The ARN of the portal, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:portal/${PortalId}

$sel:portalStartUrl:CreatePortalResponse', createPortalResponse_portalStartUrl - The URL for the IoT SiteWise Monitor portal. You can use this URL to access portals that use Amazon Web Services SSO for authentication. For portals that use IAM for authentication, you must use the IoT SiteWise console to get a URL that you can use to access the portal.

$sel:portalStatus:CreatePortalResponse', createPortalResponse_portalStatus - The status of the portal, which contains a state (CREATING after successfully calling this operation) and any error message.

$sel:ssoApplicationId:CreatePortalResponse', createPortalResponse_ssoApplicationId - The associated Amazon Web Services SSO application ID, if the portal uses Amazon Web Services SSO.

Response Lenses

createPortalResponse_httpStatus :: Lens' CreatePortalResponse Int

The response's http status code.

createPortalResponse_portalId :: Lens' CreatePortalResponse Text

The ID of the created portal.

createPortalResponse_portalArn :: Lens' CreatePortalResponse Text

The ARN of the portal, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:portal/${PortalId}

createPortalResponse_portalStartUrl :: Lens' CreatePortalResponse Text

The URL for the IoT SiteWise Monitor portal. You can use this URL to access portals that use Amazon Web Services SSO for authentication. For portals that use IAM for authentication, you must use the IoT SiteWise console to get a URL that you can use to access the portal.

createPortalResponse_portalStatus :: Lens' CreatePortalResponse PortalStatus

The status of the portal, which contains a state (CREATING after successfully calling this operation) and any error message.

createPortalResponse_ssoApplicationId :: Lens' CreatePortalResponse Text

The associated Amazon Web Services SSO application ID, if the portal uses Amazon Web Services SSO.