Network.AWS.IoTSiteWise.DeleteAssetModel

Description

Deletes an asset model. This action can't be undone. You must delete all assets created from an asset model before you can delete the model. Also, you can't delete an asset model if a parent asset model exists that contains a property formula expression that depends on the asset model that you want to delete. For more information, see Deleting assets and models in the IoT SiteWise User Guide.

Creating a Request

data DeleteAssetModel

See: newDeleteAssetModel smart constructor.

Constructors

DeleteAssetModel'

Fields

clientToken :: Maybe Text

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

assetModelId :: Text

The ID of the asset model to delete.

newDeleteAssetModel

Arguments

:: Text
$sel:assetModelId:DeleteAssetModel'

-> DeleteAssetModel

Create a value of DeleteAssetModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:DeleteAssetModel', deleteAssetModel_clientToken - A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

$sel:assetModelId:DeleteAssetModel', deleteAssetModel_assetModelId - The ID of the asset model to delete.

Request Lenses

deleteAssetModel_clientToken :: Lens' DeleteAssetModel (Maybe Text)

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

deleteAssetModel_assetModelId :: Lens' DeleteAssetModel Text

The ID of the asset model to delete.

Destructuring the Response

data DeleteAssetModelResponse

See: newDeleteAssetModelResponse smart constructor.

Constructors

DeleteAssetModelResponse'

Fields

httpStatus :: Int

The response's http status code.

assetModelStatus :: AssetModelStatus

The status of the asset model, which contains a state (DELETING after successfully calling this operation) and any error message.

newDeleteAssetModelResponse

Arguments

:: Int
$sel:httpStatus:DeleteAssetModelResponse'

-> AssetModelStatus
$sel:assetModelStatus:DeleteAssetModelResponse'

-> DeleteAssetModelResponse

Create a value of DeleteAssetModelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteAssetModelResponse', deleteAssetModelResponse_httpStatus - The response's http status code.

$sel:assetModelStatus:DeleteAssetModelResponse', deleteAssetModelResponse_assetModelStatus - The status of the asset model, which contains a state (DELETING after successfully calling this operation) and any error message.

Response Lenses

deleteAssetModelResponse_httpStatus :: Lens' DeleteAssetModelResponse Int

The response's http status code.

deleteAssetModelResponse_assetModelStatus :: Lens' DeleteAssetModelResponse AssetModelStatus

The status of the asset model, which contains a state (DELETING after successfully calling this operation) and any error message.