Network.AWS.IoTSiteWise.DescribeAccessPolicy

Description

Describes an access policy, which specifies an identity's access to an IoT SiteWise Monitor portal or project.

Creating a Request

data DescribeAccessPolicy

See: newDescribeAccessPolicy smart constructor.

Constructors

DescribeAccessPolicy'

Fields

accessPolicyId :: Text

The ID of the access policy.

newDescribeAccessPolicy

Arguments

:: Text
$sel:accessPolicyId:DescribeAccessPolicy'

-> DescribeAccessPolicy

Create a value of DescribeAccessPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessPolicyId:DescribeAccessPolicy', describeAccessPolicy_accessPolicyId - The ID of the access policy.

Request Lenses

describeAccessPolicy_accessPolicyId :: Lens' DescribeAccessPolicy Text

The ID of the access policy.

Destructuring the Response

data DescribeAccessPolicyResponse

See: newDescribeAccessPolicyResponse smart constructor.

Constructors

DescribeAccessPolicyResponse'

Fields

httpStatus :: Int

The response's http status code.

accessPolicyId :: Text

The ID of the access policy.

accessPolicyArn :: Text

The ARN of the access policy, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:access-policy/${AccessPolicyId}

accessPolicyIdentity :: Identity

The identity (Amazon Web Services SSO user, Amazon Web Services SSO group, or IAM user) to which this access policy applies.

accessPolicyResource :: Resource

The IoT SiteWise Monitor resource (portal or project) to which this access policy provides access.

accessPolicyPermission :: Permission

The access policy permission. Note that a project ADMINISTRATOR is also known as a project owner.

accessPolicyCreationDate :: POSIX

The date the access policy was created, in Unix epoch time.

accessPolicyLastUpdateDate :: POSIX

The date the access policy was last updated, in Unix epoch time.

newDescribeAccessPolicyResponse

Arguments

:: Int
$sel:httpStatus:DescribeAccessPolicyResponse'

-> Text
$sel:accessPolicyId:DescribeAccessPolicy'

-> Text
$sel:accessPolicyArn:DescribeAccessPolicyResponse'

-> Identity
$sel:accessPolicyIdentity:DescribeAccessPolicyResponse'

-> Resource
$sel:accessPolicyResource:DescribeAccessPolicyResponse'

-> Permission
$sel:accessPolicyPermission:DescribeAccessPolicyResponse'

-> UTCTime
$sel:accessPolicyCreationDate:DescribeAccessPolicyResponse'

-> UTCTime
$sel:accessPolicyLastUpdateDate:DescribeAccessPolicyResponse'

-> DescribeAccessPolicyResponse

Create a value of DescribeAccessPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeAccessPolicyResponse', describeAccessPolicyResponse_httpStatus - The response's http status code.

$sel:accessPolicyId:DescribeAccessPolicy', describeAccessPolicyResponse_accessPolicyId - The ID of the access policy.

$sel:accessPolicyArn:DescribeAccessPolicyResponse', describeAccessPolicyResponse_accessPolicyArn - The ARN of the access policy, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:access-policy/${AccessPolicyId}

$sel:accessPolicyIdentity:DescribeAccessPolicyResponse', describeAccessPolicyResponse_accessPolicyIdentity - The identity (Amazon Web Services SSO user, Amazon Web Services SSO group, or IAM user) to which this access policy applies.

$sel:accessPolicyResource:DescribeAccessPolicyResponse', describeAccessPolicyResponse_accessPolicyResource - The IoT SiteWise Monitor resource (portal or project) to which this access policy provides access.

$sel:accessPolicyPermission:DescribeAccessPolicyResponse', describeAccessPolicyResponse_accessPolicyPermission - The access policy permission. Note that a project ADMINISTRATOR is also known as a project owner.

$sel:accessPolicyCreationDate:DescribeAccessPolicyResponse', describeAccessPolicyResponse_accessPolicyCreationDate - The date the access policy was created, in Unix epoch time.

$sel:accessPolicyLastUpdateDate:DescribeAccessPolicyResponse', describeAccessPolicyResponse_accessPolicyLastUpdateDate - The date the access policy was last updated, in Unix epoch time.

Response Lenses

describeAccessPolicyResponse_httpStatus :: Lens' DescribeAccessPolicyResponse Int

The response's http status code.

describeAccessPolicyResponse_accessPolicyId :: Lens' DescribeAccessPolicyResponse Text

The ID of the access policy.

describeAccessPolicyResponse_accessPolicyArn :: Lens' DescribeAccessPolicyResponse Text

The ARN of the access policy, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:access-policy/${AccessPolicyId}

describeAccessPolicyResponse_accessPolicyIdentity :: Lens' DescribeAccessPolicyResponse Identity

The identity (Amazon Web Services SSO user, Amazon Web Services SSO group, or IAM user) to which this access policy applies.

describeAccessPolicyResponse_accessPolicyResource :: Lens' DescribeAccessPolicyResponse Resource

The IoT SiteWise Monitor resource (portal or project) to which this access policy provides access.

describeAccessPolicyResponse_accessPolicyPermission :: Lens' DescribeAccessPolicyResponse Permission

The access policy permission. Note that a project ADMINISTRATOR is also known as a project owner.

describeAccessPolicyResponse_accessPolicyCreationDate :: Lens' DescribeAccessPolicyResponse UTCTime

The date the access policy was created, in Unix epoch time.

describeAccessPolicyResponse_accessPolicyLastUpdateDate :: Lens' DescribeAccessPolicyResponse UTCTime

The date the access policy was last updated, in Unix epoch time.