Network.AWS.IoTSiteWise.DescribeAsset

Description

Retrieves information about an asset.

Creating a Request

data DescribeAsset

See: newDescribeAsset smart constructor.

Constructors

DescribeAsset'

Fields

assetId :: Text

The ID of the asset.

newDescribeAsset

Arguments

:: Text
$sel:assetId:DescribeAsset'

-> DescribeAsset

Create a value of DescribeAsset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assetId:DescribeAsset', describeAsset_assetId - The ID of the asset.

Request Lenses

describeAsset_assetId :: Lens' DescribeAsset Text

The ID of the asset.

Destructuring the Response

data DescribeAssetResponse

See: newDescribeAssetResponse smart constructor.

Constructors

DescribeAssetResponse'

Fields

assetCompositeModels :: Maybe [AssetCompositeModel]

The composite models for the asset.

httpStatus :: Int

The response's http status code.

assetId :: Text

The ID of the asset.

assetArn :: Text

The ARN of the asset, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:asset/${AssetId}

assetName :: Text

The name of the asset.

assetModelId :: Text

The ID of the asset model that was used to create the asset.

assetProperties :: [AssetProperty]

The list of asset properties for the asset.

This object doesn't include properties that you define in composite models. You can find composite model properties in the assetCompositeModels object.

assetHierarchies :: [AssetHierarchy]

A list of asset hierarchies that each contain a hierarchyId. A hierarchy specifies allowed parent/child asset relationships.

assetCreationDate :: POSIX

The date the asset was created, in Unix epoch time.

assetLastUpdateDate :: POSIX

The date the asset was last updated, in Unix epoch time.

assetStatus :: AssetStatus

The current status of the asset, which contains a state and any error message.

newDescribeAssetResponse

Arguments

:: Int
$sel:httpStatus:DescribeAssetResponse'

-> Text
$sel:assetId:DescribeAsset'

-> Text
$sel:assetArn:DescribeAssetResponse'

-> Text
$sel:assetName:DescribeAssetResponse'

-> Text
$sel:assetModelId:DescribeAssetResponse'

-> UTCTime
$sel:assetCreationDate:DescribeAssetResponse'

-> UTCTime
$sel:assetLastUpdateDate:DescribeAssetResponse'

-> AssetStatus
$sel:assetStatus:DescribeAssetResponse'

-> DescribeAssetResponse

Create a value of DescribeAssetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assetCompositeModels:DescribeAssetResponse', describeAssetResponse_assetCompositeModels - The composite models for the asset.

$sel:httpStatus:DescribeAssetResponse', describeAssetResponse_httpStatus - The response's http status code.

$sel:assetId:DescribeAsset', describeAssetResponse_assetId - The ID of the asset.

$sel:assetArn:DescribeAssetResponse', describeAssetResponse_assetArn - The ARN of the asset, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:asset/${AssetId}

$sel:assetName:DescribeAssetResponse', describeAssetResponse_assetName - The name of the asset.

$sel:assetModelId:DescribeAssetResponse', describeAssetResponse_assetModelId - The ID of the asset model that was used to create the asset.

$sel:assetProperties:DescribeAssetResponse', describeAssetResponse_assetProperties - The list of asset properties for the asset.

This object doesn't include properties that you define in composite models. You can find composite model properties in the assetCompositeModels object.

$sel:assetHierarchies:DescribeAssetResponse', describeAssetResponse_assetHierarchies - A list of asset hierarchies that each contain a hierarchyId. A hierarchy specifies allowed parent/child asset relationships.

$sel:assetCreationDate:DescribeAssetResponse', describeAssetResponse_assetCreationDate - The date the asset was created, in Unix epoch time.

$sel:assetLastUpdateDate:DescribeAssetResponse', describeAssetResponse_assetLastUpdateDate - The date the asset was last updated, in Unix epoch time.

$sel:assetStatus:DescribeAssetResponse', describeAssetResponse_assetStatus - The current status of the asset, which contains a state and any error message.

Response Lenses

describeAssetResponse_assetCompositeModels :: Lens' DescribeAssetResponse (Maybe [AssetCompositeModel])

The composite models for the asset.

describeAssetResponse_httpStatus :: Lens' DescribeAssetResponse Int

The response's http status code.

describeAssetResponse_assetId :: Lens' DescribeAssetResponse Text

The ID of the asset.

describeAssetResponse_assetArn :: Lens' DescribeAssetResponse Text

The ARN of the asset, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:asset/${AssetId}

describeAssetResponse_assetName :: Lens' DescribeAssetResponse Text

The name of the asset.

describeAssetResponse_assetModelId :: Lens' DescribeAssetResponse Text

The ID of the asset model that was used to create the asset.

describeAssetResponse_assetProperties :: Lens' DescribeAssetResponse [AssetProperty]

The list of asset properties for the asset.

This object doesn't include properties that you define in composite models. You can find composite model properties in the assetCompositeModels object.

describeAssetResponse_assetHierarchies :: Lens' DescribeAssetResponse [AssetHierarchy]

A list of asset hierarchies that each contain a hierarchyId. A hierarchy specifies allowed parent/child asset relationships.

describeAssetResponse_assetCreationDate :: Lens' DescribeAssetResponse UTCTime

The date the asset was created, in Unix epoch time.

describeAssetResponse_assetLastUpdateDate :: Lens' DescribeAssetResponse UTCTime

The date the asset was last updated, in Unix epoch time.

describeAssetResponse_assetStatus :: Lens' DescribeAssetResponse AssetStatus

The current status of the asset, which contains a state and any error message.