Network.AWS.IoTSiteWise.DescribeDashboard

Description

Retrieves information about a dashboard.

Creating a Request

data DescribeDashboard

See: newDescribeDashboard smart constructor.

Constructors

DescribeDashboard'

Fields

dashboardId :: Text

The ID of the dashboard.

newDescribeDashboard

Arguments

:: Text
$sel:dashboardId:DescribeDashboard'

-> DescribeDashboard

Create a value of DescribeDashboard with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dashboardId:DescribeDashboard', describeDashboard_dashboardId - The ID of the dashboard.

Request Lenses

describeDashboard_dashboardId :: Lens' DescribeDashboard Text

The ID of the dashboard.

Destructuring the Response

data DescribeDashboardResponse

See: newDescribeDashboardResponse smart constructor.

Constructors

DescribeDashboardResponse'

Fields

dashboardDescription :: Maybe Text

The dashboard's description.

httpStatus :: Int

The response's http status code.

dashboardId :: Text

The ID of the dashboard.

dashboardArn :: Text

The ARN of the dashboard, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:dashboard/${DashboardId}

dashboardName :: Text

The name of the dashboard.

projectId :: Text

The ID of the project that the dashboard is in.

dashboardDefinition :: Text

The dashboard's definition JSON literal. For detailed information, see Creating dashboards (CLI) in the IoT SiteWise User Guide.

dashboardCreationDate :: POSIX

The date the dashboard was created, in Unix epoch time.

dashboardLastUpdateDate :: POSIX

The date the dashboard was last updated, in Unix epoch time.

newDescribeDashboardResponse

Arguments

:: Int
$sel:httpStatus:DescribeDashboardResponse'

-> Text
$sel:dashboardId:DescribeDashboard'

-> Text
$sel:dashboardArn:DescribeDashboardResponse'

-> Text
$sel:dashboardName:DescribeDashboardResponse'

-> Text
$sel:projectId:DescribeDashboardResponse'

-> Text
$sel:dashboardDefinition:DescribeDashboardResponse'

-> UTCTime
$sel:dashboardCreationDate:DescribeDashboardResponse'

-> UTCTime
$sel:dashboardLastUpdateDate:DescribeDashboardResponse'

-> DescribeDashboardResponse

Create a value of DescribeDashboardResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dashboardDescription:DescribeDashboardResponse', describeDashboardResponse_dashboardDescription - The dashboard's description.

$sel:httpStatus:DescribeDashboardResponse', describeDashboardResponse_httpStatus - The response's http status code.

$sel:dashboardId:DescribeDashboard', describeDashboardResponse_dashboardId - The ID of the dashboard.

$sel:dashboardArn:DescribeDashboardResponse', describeDashboardResponse_dashboardArn - The ARN of the dashboard, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:dashboard/${DashboardId}

$sel:dashboardName:DescribeDashboardResponse', describeDashboardResponse_dashboardName - The name of the dashboard.

$sel:projectId:DescribeDashboardResponse', describeDashboardResponse_projectId - The ID of the project that the dashboard is in.

$sel:dashboardDefinition:DescribeDashboardResponse', describeDashboardResponse_dashboardDefinition - The dashboard's definition JSON literal. For detailed information, see Creating dashboards (CLI) in the IoT SiteWise User Guide.

$sel:dashboardCreationDate:DescribeDashboardResponse', describeDashboardResponse_dashboardCreationDate - The date the dashboard was created, in Unix epoch time.

$sel:dashboardLastUpdateDate:DescribeDashboardResponse', describeDashboardResponse_dashboardLastUpdateDate - The date the dashboard was last updated, in Unix epoch time.

Response Lenses

describeDashboardResponse_dashboardDescription :: Lens' DescribeDashboardResponse (Maybe Text)

The dashboard's description.

describeDashboardResponse_httpStatus :: Lens' DescribeDashboardResponse Int

The response's http status code.

describeDashboardResponse_dashboardId :: Lens' DescribeDashboardResponse Text

The ID of the dashboard.

describeDashboardResponse_dashboardArn :: Lens' DescribeDashboardResponse Text

The ARN of the dashboard, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:dashboard/${DashboardId}

describeDashboardResponse_dashboardName :: Lens' DescribeDashboardResponse Text

The name of the dashboard.

describeDashboardResponse_projectId :: Lens' DescribeDashboardResponse Text

The ID of the project that the dashboard is in.

describeDashboardResponse_dashboardDefinition :: Lens' DescribeDashboardResponse Text

The dashboard's definition JSON literal. For detailed information, see Creating dashboards (CLI) in the IoT SiteWise User Guide.

describeDashboardResponse_dashboardCreationDate :: Lens' DescribeDashboardResponse UTCTime

The date the dashboard was created, in Unix epoch time.

describeDashboardResponse_dashboardLastUpdateDate :: Lens' DescribeDashboardResponse UTCTime

The date the dashboard was last updated, in Unix epoch time.