Network.AWS.IoTSiteWise.DescribePortal

Description

Retrieves information about a portal.

Creating a Request

data DescribePortal

See: newDescribePortal smart constructor.

Constructors

DescribePortal'

Fields

portalId :: Text

The ID of the portal.

newDescribePortal

Arguments

:: Text
$sel:portalId:DescribePortal'

-> DescribePortal

Create a value of DescribePortal with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:portalId:DescribePortal', describePortal_portalId - The ID of the portal.

Request Lenses

describePortal_portalId :: Lens' DescribePortal Text

The ID of the portal.

Destructuring the Response

data DescribePortalResponse

See: newDescribePortalResponse smart constructor.

Constructors

DescribePortalResponse'

Fields

portalAuthMode :: Maybe AuthMode

The service to use to authenticate users to the portal.

portalDescription :: Maybe Text

The portal's description.

notificationSenderEmail :: Maybe Text

The email address that sends alarm notifications.

portalLogoImageLocation :: Maybe ImageLocation

The portal's logo image, which is available at a URL.

alarms :: Maybe Alarms

Contains the configuration information of an alarm created in an IoT SiteWise Monitor portal.

roleArn :: Maybe Text

The ARN of the service role that allows the portal's users to access your IoT SiteWise resources on your behalf. For more information, see Using service roles for IoT SiteWise Monitor in the IoT SiteWise User Guide.

httpStatus :: Int

The response's http status code.

portalId :: Text

The ID of the portal.

portalArn :: Text

The ARN of the portal, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:portal/${PortalId}

portalName :: Text

The name of the portal.

portalClientId :: Text

The Amazon Web Services SSO application generated client ID (used with Amazon Web Services SSO APIs). IoT SiteWise includes portalClientId for only portals that use Amazon Web Services SSO to authenticate users.

portalStartUrl :: Text

The URL for the IoT SiteWise Monitor portal. You can use this URL to access portals that use Amazon Web Services SSO for authentication. For portals that use IAM for authentication, you must use the IoT SiteWise console to get a URL that you can use to access the portal.

portalContactEmail :: Text

The Amazon Web Services administrator's contact email address.

portalStatus :: PortalStatus

The current status of the portal, which contains a state and any error message.

portalCreationDate :: POSIX

The date the portal was created, in Unix epoch time.

portalLastUpdateDate :: POSIX

The date the portal was last updated, in Unix epoch time.

newDescribePortalResponse

Arguments

:: Int
$sel:httpStatus:DescribePortalResponse'

-> Text
$sel:portalId:DescribePortal'

-> Text
$sel:portalArn:DescribePortalResponse'

-> Text
$sel:portalName:DescribePortalResponse'

-> Text
$sel:portalClientId:DescribePortalResponse'

-> Text
$sel:portalStartUrl:DescribePortalResponse'

-> Text
$sel:portalContactEmail:DescribePortalResponse'

-> PortalStatus
$sel:portalStatus:DescribePortalResponse'

-> UTCTime
$sel:portalCreationDate:DescribePortalResponse'

-> UTCTime
$sel:portalLastUpdateDate:DescribePortalResponse'

-> DescribePortalResponse

Create a value of DescribePortalResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:portalAuthMode:DescribePortalResponse', describePortalResponse_portalAuthMode - The service to use to authenticate users to the portal.

$sel:portalDescription:DescribePortalResponse', describePortalResponse_portalDescription - The portal's description.

$sel:notificationSenderEmail:DescribePortalResponse', describePortalResponse_notificationSenderEmail - The email address that sends alarm notifications.

$sel:portalLogoImageLocation:DescribePortalResponse', describePortalResponse_portalLogoImageLocation - The portal's logo image, which is available at a URL.

$sel:alarms:DescribePortalResponse', describePortalResponse_alarms - Contains the configuration information of an alarm created in an IoT SiteWise Monitor portal.

$sel:roleArn:DescribePortalResponse', describePortalResponse_roleArn - The ARN of the service role that allows the portal's users to access your IoT SiteWise resources on your behalf. For more information, see Using service roles for IoT SiteWise Monitor in the IoT SiteWise User Guide.

$sel:httpStatus:DescribePortalResponse', describePortalResponse_httpStatus - The response's http status code.

$sel:portalId:DescribePortal', describePortalResponse_portalId - The ID of the portal.

$sel:portalArn:DescribePortalResponse', describePortalResponse_portalArn - The ARN of the portal, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:portal/${PortalId}

$sel:portalName:DescribePortalResponse', describePortalResponse_portalName - The name of the portal.

$sel:portalClientId:DescribePortalResponse', describePortalResponse_portalClientId - The Amazon Web Services SSO application generated client ID (used with Amazon Web Services SSO APIs). IoT SiteWise includes portalClientId for only portals that use Amazon Web Services SSO to authenticate users.

$sel:portalStartUrl:DescribePortalResponse', describePortalResponse_portalStartUrl - The URL for the IoT SiteWise Monitor portal. You can use this URL to access portals that use Amazon Web Services SSO for authentication. For portals that use IAM for authentication, you must use the IoT SiteWise console to get a URL that you can use to access the portal.

$sel:portalContactEmail:DescribePortalResponse', describePortalResponse_portalContactEmail - The Amazon Web Services administrator's contact email address.

$sel:portalStatus:DescribePortalResponse', describePortalResponse_portalStatus - The current status of the portal, which contains a state and any error message.

$sel:portalCreationDate:DescribePortalResponse', describePortalResponse_portalCreationDate - The date the portal was created, in Unix epoch time.

$sel:portalLastUpdateDate:DescribePortalResponse', describePortalResponse_portalLastUpdateDate - The date the portal was last updated, in Unix epoch time.

Response Lenses

describePortalResponse_portalAuthMode :: Lens' DescribePortalResponse (Maybe AuthMode)

The service to use to authenticate users to the portal.

describePortalResponse_portalDescription :: Lens' DescribePortalResponse (Maybe Text)

The portal's description.

describePortalResponse_notificationSenderEmail :: Lens' DescribePortalResponse (Maybe Text)

The email address that sends alarm notifications.

describePortalResponse_portalLogoImageLocation :: Lens' DescribePortalResponse (Maybe ImageLocation)

The portal's logo image, which is available at a URL.

describePortalResponse_alarms :: Lens' DescribePortalResponse (Maybe Alarms)

Contains the configuration information of an alarm created in an IoT SiteWise Monitor portal.

describePortalResponse_roleArn :: Lens' DescribePortalResponse (Maybe Text)

The ARN of the service role that allows the portal's users to access your IoT SiteWise resources on your behalf. For more information, see Using service roles for IoT SiteWise Monitor in the IoT SiteWise User Guide.

describePortalResponse_httpStatus :: Lens' DescribePortalResponse Int

The response's http status code.

describePortalResponse_portalId :: Lens' DescribePortalResponse Text

The ID of the portal.

describePortalResponse_portalArn :: Lens' DescribePortalResponse Text

The ARN of the portal, which has the following format.

arn:${Partition}:iotsitewise:${Region}:${Account}:portal/${PortalId}

describePortalResponse_portalName :: Lens' DescribePortalResponse Text

The name of the portal.

describePortalResponse_portalClientId :: Lens' DescribePortalResponse Text

The Amazon Web Services SSO application generated client ID (used with Amazon Web Services SSO APIs). IoT SiteWise includes portalClientId for only portals that use Amazon Web Services SSO to authenticate users.

describePortalResponse_portalStartUrl :: Lens' DescribePortalResponse Text

The URL for the IoT SiteWise Monitor portal. You can use this URL to access portals that use Amazon Web Services SSO for authentication. For portals that use IAM for authentication, you must use the IoT SiteWise console to get a URL that you can use to access the portal.

describePortalResponse_portalContactEmail :: Lens' DescribePortalResponse Text

The Amazon Web Services administrator's contact email address.

describePortalResponse_portalStatus :: Lens' DescribePortalResponse PortalStatus

The current status of the portal, which contains a state and any error message.

describePortalResponse_portalCreationDate :: Lens' DescribePortalResponse UTCTime

The date the portal was created, in Unix epoch time.

describePortalResponse_portalLastUpdateDate :: Lens' DescribePortalResponse UTCTime

The date the portal was last updated, in Unix epoch time.