Network.AWS.IoTSiteWise.DescribeStorageConfiguration

Description

Retrieves information about the storage configuration for IoT SiteWise.

Creating a Request

data DescribeStorageConfiguration

See: newDescribeStorageConfiguration smart constructor.

Constructors

DescribeStorageConfiguration'

newDescribeStorageConfiguration :: DescribeStorageConfiguration

Create a value of DescribeStorageConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data DescribeStorageConfigurationResponse

See: newDescribeStorageConfigurationResponse smart constructor.

Constructors

DescribeStorageConfigurationResponse'

Fields

multiLayerStorage :: Maybe MultiLayerStorage

Contains information about the storage destination.

lastUpdateDate :: Maybe POSIX

The date the storage configuration was last updated, in Unix epoch time.

httpStatus :: Int

The response's http status code.

storageType :: StorageType

The type of storage that you specified for your data. The storage type can be one of the following values:

SITEWISE_DEFAULT_STORAGE – IoT SiteWise replicates your data into a service managed database.

MULTI_LAYER_STORAGE – IoT SiteWise replicates your data into a service managed database and saves a copy of your raw data and metadata in an Amazon S3 object that you specified.

configurationStatus :: ConfigurationStatus

newDescribeStorageConfigurationResponse

Arguments

:: Int
$sel:httpStatus:DescribeStorageConfigurationResponse'

-> StorageType
$sel:storageType:DescribeStorageConfigurationResponse'

-> ConfigurationStatus
$sel:configurationStatus:DescribeStorageConfigurationResponse'

-> DescribeStorageConfigurationResponse

Create a value of DescribeStorageConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:multiLayerStorage:DescribeStorageConfigurationResponse', describeStorageConfigurationResponse_multiLayerStorage - Contains information about the storage destination.

$sel:lastUpdateDate:DescribeStorageConfigurationResponse', describeStorageConfigurationResponse_lastUpdateDate - The date the storage configuration was last updated, in Unix epoch time.

$sel:httpStatus:DescribeStorageConfigurationResponse', describeStorageConfigurationResponse_httpStatus - The response's http status code.

$sel:storageType:DescribeStorageConfigurationResponse', describeStorageConfigurationResponse_storageType - The type of storage that you specified for your data. The storage type can be one of the following values:

SITEWISE_DEFAULT_STORAGE – IoT SiteWise replicates your data into a service managed database.

MULTI_LAYER_STORAGE – IoT SiteWise replicates your data into a service managed database and saves a copy of your raw data and metadata in an Amazon S3 object that you specified.

$sel:configurationStatus:DescribeStorageConfigurationResponse', describeStorageConfigurationResponse_configurationStatus - Undocumented member.

Response Lenses

describeStorageConfigurationResponse_multiLayerStorage :: Lens' DescribeStorageConfigurationResponse (Maybe MultiLayerStorage)

Contains information about the storage destination.

describeStorageConfigurationResponse_lastUpdateDate :: Lens' DescribeStorageConfigurationResponse (Maybe UTCTime)

The date the storage configuration was last updated, in Unix epoch time.

describeStorageConfigurationResponse_httpStatus :: Lens' DescribeStorageConfigurationResponse Int

The response's http status code.

describeStorageConfigurationResponse_storageType :: Lens' DescribeStorageConfigurationResponse StorageType

The type of storage that you specified for your data. The storage type can be one of the following values:

SITEWISE_DEFAULT_STORAGE – IoT SiteWise replicates your data into a service managed database.

MULTI_LAYER_STORAGE – IoT SiteWise replicates your data into a service managed database and saves a copy of your raw data and metadata in an Amazon S3 object that you specified.

describeStorageConfigurationResponse_configurationStatus :: Lens' DescribeStorageConfigurationResponse ConfigurationStatus

Undocumented member.