Network.AWS.IoTSiteWise.GetAssetPropertyValue

Description

Gets an asset property's current value. For more information, see Querying current values in the IoT SiteWise User Guide.

To identify an asset property, you must specify one of the following:

The assetId and propertyId of an asset property.

A propertyAlias, which is a data stream alias (for example, /company/windfarm/3/turbine/7/temperature). To define an asset property's alias, see UpdateAssetProperty.

Creating a Request

data GetAssetPropertyValue

See: newGetAssetPropertyValue smart constructor.

Constructors

GetAssetPropertyValue'

Fields

propertyAlias :: Maybe Text

The alias that identifies the property, such as an OPC-UA server data stream path (for example, /company/windfarm/3/turbine/7/temperature). For more information, see Mapping industrial data streams to asset properties in the IoT SiteWise User Guide.

propertyId :: Maybe Text

The ID of the asset property.

assetId :: Maybe Text

The ID of the asset.

newGetAssetPropertyValue :: GetAssetPropertyValue

Create a value of GetAssetPropertyValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:propertyAlias:GetAssetPropertyValue', getAssetPropertyValue_propertyAlias - The alias that identifies the property, such as an OPC-UA server data stream path (for example, /company/windfarm/3/turbine/7/temperature). For more information, see Mapping industrial data streams to asset properties in the IoT SiteWise User Guide.

$sel:propertyId:GetAssetPropertyValue', getAssetPropertyValue_propertyId - The ID of the asset property.

$sel:assetId:GetAssetPropertyValue', getAssetPropertyValue_assetId - The ID of the asset.

Request Lenses

getAssetPropertyValue_propertyAlias :: Lens' GetAssetPropertyValue (Maybe Text)

The alias that identifies the property, such as an OPC-UA server data stream path (for example, /company/windfarm/3/turbine/7/temperature). For more information, see Mapping industrial data streams to asset properties in the IoT SiteWise User Guide.

getAssetPropertyValue_propertyId :: Lens' GetAssetPropertyValue (Maybe Text)

The ID of the asset property.

getAssetPropertyValue_assetId :: Lens' GetAssetPropertyValue (Maybe Text)

The ID of the asset.

Destructuring the Response

data GetAssetPropertyValueResponse

See: newGetAssetPropertyValueResponse smart constructor.

Constructors

GetAssetPropertyValueResponse'

Fields

propertyValue :: Maybe AssetPropertyValue

The current asset property value.

httpStatus :: Int

The response's http status code.

newGetAssetPropertyValueResponse

Arguments

:: Int
$sel:httpStatus:GetAssetPropertyValueResponse'

-> GetAssetPropertyValueResponse

Create a value of GetAssetPropertyValueResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:propertyValue:GetAssetPropertyValueResponse', getAssetPropertyValueResponse_propertyValue - The current asset property value.

$sel:httpStatus:GetAssetPropertyValueResponse', getAssetPropertyValueResponse_httpStatus - The response's http status code.

Response Lenses

getAssetPropertyValueResponse_propertyValue :: Lens' GetAssetPropertyValueResponse (Maybe AssetPropertyValue)

The current asset property value.

getAssetPropertyValueResponse_httpStatus :: Lens' GetAssetPropertyValueResponse Int

The response's http status code.