Network.AWS.IoTSiteWise.GetAssetPropertyValueHistory

Description

Gets the history of an asset property's values. For more information, see Querying historical values in the IoT SiteWise User Guide.

To identify an asset property, you must specify one of the following:

The assetId and propertyId of an asset property.

A propertyAlias, which is a data stream alias (for example, /company/windfarm/3/turbine/7/temperature). To define an asset property's alias, see UpdateAssetProperty.

This operation returns paginated results.

Creating a Request

data GetAssetPropertyValueHistory

See: newGetAssetPropertyValueHistory smart constructor.

Constructors

GetAssetPropertyValueHistory'

Fields

endDate :: Maybe POSIX

The inclusive end of the range from which to query historical data, expressed in seconds in Unix epoch time.

qualities :: Maybe (NonEmpty Quality)

The quality by which to filter asset data.

timeOrdering :: Maybe TimeOrdering

The chronological sorting order of the requested information.

Default: ASCENDING

propertyAlias :: Maybe Text

The alias that identifies the property, such as an OPC-UA server data stream path (for example, /company/windfarm/3/turbine/7/temperature). For more information, see Mapping industrial data streams to asset properties in the IoT SiteWise User Guide.

startDate :: Maybe POSIX

The exclusive start of the range from which to query historical data, expressed in seconds in Unix epoch time.

nextToken :: Maybe Text

The token to be used for the next set of paginated results.

propertyId :: Maybe Text

The ID of the asset property.

assetId :: Maybe Text

The ID of the asset.

maxResults :: Maybe Natural

The maximum number of results to return for each paginated request.

Default: 100

newGetAssetPropertyValueHistory :: GetAssetPropertyValueHistory

Create a value of GetAssetPropertyValueHistory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endDate:GetAssetPropertyValueHistory', getAssetPropertyValueHistory_endDate - The inclusive end of the range from which to query historical data, expressed in seconds in Unix epoch time.

$sel:qualities:GetAssetPropertyValueHistory', getAssetPropertyValueHistory_qualities - The quality by which to filter asset data.

$sel:timeOrdering:GetAssetPropertyValueHistory', getAssetPropertyValueHistory_timeOrdering - The chronological sorting order of the requested information.

Default: ASCENDING

$sel:propertyAlias:GetAssetPropertyValueHistory', getAssetPropertyValueHistory_propertyAlias - The alias that identifies the property, such as an OPC-UA server data stream path (for example, /company/windfarm/3/turbine/7/temperature). For more information, see Mapping industrial data streams to asset properties in the IoT SiteWise User Guide.

$sel:startDate:GetAssetPropertyValueHistory', getAssetPropertyValueHistory_startDate - The exclusive start of the range from which to query historical data, expressed in seconds in Unix epoch time.

$sel:nextToken:GetAssetPropertyValueHistory', getAssetPropertyValueHistory_nextToken - The token to be used for the next set of paginated results.

$sel:propertyId:GetAssetPropertyValueHistory', getAssetPropertyValueHistory_propertyId - The ID of the asset property.

$sel:assetId:GetAssetPropertyValueHistory', getAssetPropertyValueHistory_assetId - The ID of the asset.

$sel:maxResults:GetAssetPropertyValueHistory', getAssetPropertyValueHistory_maxResults - The maximum number of results to return for each paginated request.

Default: 100

Request Lenses

getAssetPropertyValueHistory_endDate :: Lens' GetAssetPropertyValueHistory (Maybe UTCTime)

The inclusive end of the range from which to query historical data, expressed in seconds in Unix epoch time.

getAssetPropertyValueHistory_qualities :: Lens' GetAssetPropertyValueHistory (Maybe (NonEmpty Quality))

The quality by which to filter asset data.

getAssetPropertyValueHistory_timeOrdering :: Lens' GetAssetPropertyValueHistory (Maybe TimeOrdering)

The chronological sorting order of the requested information.

Default: ASCENDING

getAssetPropertyValueHistory_propertyAlias :: Lens' GetAssetPropertyValueHistory (Maybe Text)

The alias that identifies the property, such as an OPC-UA server data stream path (for example, /company/windfarm/3/turbine/7/temperature). For more information, see Mapping industrial data streams to asset properties in the IoT SiteWise User Guide.

getAssetPropertyValueHistory_startDate :: Lens' GetAssetPropertyValueHistory (Maybe UTCTime)

The exclusive start of the range from which to query historical data, expressed in seconds in Unix epoch time.

getAssetPropertyValueHistory_nextToken :: Lens' GetAssetPropertyValueHistory (Maybe Text)

The token to be used for the next set of paginated results.

getAssetPropertyValueHistory_propertyId :: Lens' GetAssetPropertyValueHistory (Maybe Text)

The ID of the asset property.

getAssetPropertyValueHistory_assetId :: Lens' GetAssetPropertyValueHistory (Maybe Text)

The ID of the asset.

getAssetPropertyValueHistory_maxResults :: Lens' GetAssetPropertyValueHistory (Maybe Natural)

The maximum number of results to return for each paginated request.

Default: 100

Destructuring the Response

data GetAssetPropertyValueHistoryResponse

See: newGetAssetPropertyValueHistoryResponse smart constructor.

Constructors

GetAssetPropertyValueHistoryResponse'

Fields

nextToken :: Maybe Text

The token for the next set of results, or null if there are no additional results.

httpStatus :: Int

The response's http status code.

assetPropertyValueHistory :: [AssetPropertyValue]

The asset property's value history.

newGetAssetPropertyValueHistoryResponse

Arguments

:: Int
$sel:httpStatus:GetAssetPropertyValueHistoryResponse'

-> GetAssetPropertyValueHistoryResponse

Create a value of GetAssetPropertyValueHistoryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetAssetPropertyValueHistory', getAssetPropertyValueHistoryResponse_nextToken - The token for the next set of results, or null if there are no additional results.

$sel:httpStatus:GetAssetPropertyValueHistoryResponse', getAssetPropertyValueHistoryResponse_httpStatus - The response's http status code.

$sel:assetPropertyValueHistory:GetAssetPropertyValueHistoryResponse', getAssetPropertyValueHistoryResponse_assetPropertyValueHistory - The asset property's value history.

Response Lenses

getAssetPropertyValueHistoryResponse_nextToken :: Lens' GetAssetPropertyValueHistoryResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

getAssetPropertyValueHistoryResponse_httpStatus :: Lens' GetAssetPropertyValueHistoryResponse Int

The response's http status code.

getAssetPropertyValueHistoryResponse_assetPropertyValueHistory :: Lens' GetAssetPropertyValueHistoryResponse [AssetPropertyValue]

The asset property's value history.