Network.AWS.IoTSiteWise.ListAssetModels

Description

Retrieves a paginated list of summaries of all asset models.

This operation returns paginated results.

Creating a Request

data ListAssetModels

See: newListAssetModels smart constructor.

Constructors

ListAssetModels'

Fields

nextToken :: Maybe Text

The token to be used for the next set of paginated results.

maxResults :: Maybe Natural

The maximum number of results to return for each paginated request.

Default: 50

newListAssetModels :: ListAssetModels

Create a value of ListAssetModels with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAssetModels', listAssetModels_nextToken - The token to be used for the next set of paginated results.

$sel:maxResults:ListAssetModels', listAssetModels_maxResults - The maximum number of results to return for each paginated request.

Default: 50

Request Lenses

listAssetModels_nextToken :: Lens' ListAssetModels (Maybe Text)

The token to be used for the next set of paginated results.

listAssetModels_maxResults :: Lens' ListAssetModels (Maybe Natural)

The maximum number of results to return for each paginated request.

Default: 50

Destructuring the Response

data ListAssetModelsResponse

See: newListAssetModelsResponse smart constructor.

Constructors

ListAssetModelsResponse'

Fields

nextToken :: Maybe Text

The token for the next set of results, or null if there are no additional results.

httpStatus :: Int

The response's http status code.

assetModelSummaries :: [AssetModelSummary]

A list that summarizes each asset model.

newListAssetModelsResponse

Arguments

:: Int
$sel:httpStatus:ListAssetModelsResponse'

-> ListAssetModelsResponse

Create a value of ListAssetModelsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAssetModels', listAssetModelsResponse_nextToken - The token for the next set of results, or null if there are no additional results.

$sel:httpStatus:ListAssetModelsResponse', listAssetModelsResponse_httpStatus - The response's http status code.

$sel:assetModelSummaries:ListAssetModelsResponse', listAssetModelsResponse_assetModelSummaries - A list that summarizes each asset model.

Response Lenses

listAssetModelsResponse_nextToken :: Lens' ListAssetModelsResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

listAssetModelsResponse_httpStatus :: Lens' ListAssetModelsResponse Int

The response's http status code.

listAssetModelsResponse_assetModelSummaries :: Lens' ListAssetModelsResponse [AssetModelSummary]

A list that summarizes each asset model.