Network.AWS.IoTSiteWise.ListProjectAssets

Description

Retrieves a paginated list of assets associated with an IoT SiteWise Monitor project.

This operation returns paginated results.

Creating a Request

data ListProjectAssets

See: newListProjectAssets smart constructor.

Constructors

ListProjectAssets'

Fields

nextToken :: Maybe Text

The token to be used for the next set of paginated results.

maxResults :: Maybe Natural

The maximum number of results to return for each paginated request.

Default: 50

projectId :: Text

The ID of the project.

newListProjectAssets

Arguments

:: Text
$sel:projectId:ListProjectAssets'

-> ListProjectAssets

Create a value of ListProjectAssets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListProjectAssets', listProjectAssets_nextToken - The token to be used for the next set of paginated results.

$sel:maxResults:ListProjectAssets', listProjectAssets_maxResults - The maximum number of results to return for each paginated request.

Default: 50

$sel:projectId:ListProjectAssets', listProjectAssets_projectId - The ID of the project.

Request Lenses

listProjectAssets_nextToken :: Lens' ListProjectAssets (Maybe Text)

The token to be used for the next set of paginated results.

listProjectAssets_maxResults :: Lens' ListProjectAssets (Maybe Natural)

The maximum number of results to return for each paginated request.

Default: 50

listProjectAssets_projectId :: Lens' ListProjectAssets Text

The ID of the project.

Destructuring the Response

data ListProjectAssetsResponse

See: newListProjectAssetsResponse smart constructor.

Constructors

ListProjectAssetsResponse'

Fields

nextToken :: Maybe Text

The token for the next set of results, or null if there are no additional results.

httpStatus :: Int

The response's http status code.

assetIds :: [Text]

A list that contains the IDs of each asset associated with the project.

newListProjectAssetsResponse

Arguments

:: Int
$sel:httpStatus:ListProjectAssetsResponse'

-> ListProjectAssetsResponse

Create a value of ListProjectAssetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListProjectAssets', listProjectAssetsResponse_nextToken - The token for the next set of results, or null if there are no additional results.

$sel:httpStatus:ListProjectAssetsResponse', listProjectAssetsResponse_httpStatus - The response's http status code.

$sel:assetIds:ListProjectAssetsResponse', listProjectAssetsResponse_assetIds - A list that contains the IDs of each asset associated with the project.

Response Lenses

listProjectAssetsResponse_nextToken :: Lens' ListProjectAssetsResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

listProjectAssetsResponse_httpStatus :: Lens' ListProjectAssetsResponse Int

The response's http status code.

listProjectAssetsResponse_assetIds :: Lens' ListProjectAssetsResponse [Text]

A list that contains the IDs of each asset associated with the project.