Network.AWS.IoTSiteWise.PutStorageConfiguration

Description

Configures storage settings for IoT SiteWise.

Creating a Request

data PutStorageConfiguration

See: newPutStorageConfiguration smart constructor.

Constructors

PutStorageConfiguration'

Fields

multiLayerStorage :: Maybe MultiLayerStorage

Identifies a storage destination. If you specified MULTI_LAYER_STORAGE for the storage type, you must specify a MultiLayerStorage object.

storageType :: StorageType

The type of storage that you specified for your data. The storage type can be one of the following values:

SITEWISE_DEFAULT_STORAGE – IoT SiteWise replicates your data into a service managed database.

MULTI_LAYER_STORAGE – IoT SiteWise replicates your data into a service managed database and saves a copy of your raw data and metadata in an Amazon S3 object that you specified.

newPutStorageConfiguration

Arguments

:: StorageType
$sel:storageType:PutStorageConfiguration'

-> PutStorageConfiguration

Create a value of PutStorageConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:multiLayerStorage:PutStorageConfiguration', putStorageConfiguration_multiLayerStorage - Identifies a storage destination. If you specified MULTI_LAYER_STORAGE for the storage type, you must specify a MultiLayerStorage object.

$sel:storageType:PutStorageConfiguration', putStorageConfiguration_storageType - The type of storage that you specified for your data. The storage type can be one of the following values:

SITEWISE_DEFAULT_STORAGE – IoT SiteWise replicates your data into a service managed database.

MULTI_LAYER_STORAGE – IoT SiteWise replicates your data into a service managed database and saves a copy of your raw data and metadata in an Amazon S3 object that you specified.

Request Lenses

putStorageConfiguration_multiLayerStorage :: Lens' PutStorageConfiguration (Maybe MultiLayerStorage)

Identifies a storage destination. If you specified MULTI_LAYER_STORAGE for the storage type, you must specify a MultiLayerStorage object.

putStorageConfiguration_storageType :: Lens' PutStorageConfiguration StorageType

The type of storage that you specified for your data. The storage type can be one of the following values:

SITEWISE_DEFAULT_STORAGE – IoT SiteWise replicates your data into a service managed database.

MULTI_LAYER_STORAGE – IoT SiteWise replicates your data into a service managed database and saves a copy of your raw data and metadata in an Amazon S3 object that you specified.

Destructuring the Response

data PutStorageConfigurationResponse

See: newPutStorageConfigurationResponse smart constructor.

Constructors

PutStorageConfigurationResponse'

Fields

multiLayerStorage :: Maybe MultiLayerStorage

Contains information about the storage destination.

httpStatus :: Int

The response's http status code.

storageType :: StorageType

The type of storage that you specified for your data. The storage type can be one of the following values:

SITEWISE_DEFAULT_STORAGE – IoT SiteWise replicates your data into a service managed database.

MULTI_LAYER_STORAGE – IoT SiteWise replicates your data into a service managed database and saves a copy of your raw data and metadata in an Amazon S3 object that you specified.

configurationStatus :: ConfigurationStatus

newPutStorageConfigurationResponse

Arguments

:: Int
$sel:httpStatus:PutStorageConfigurationResponse'

-> StorageType
$sel:storageType:PutStorageConfiguration'

-> ConfigurationStatus
$sel:configurationStatus:PutStorageConfigurationResponse'

-> PutStorageConfigurationResponse

Create a value of PutStorageConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:multiLayerStorage:PutStorageConfiguration', putStorageConfigurationResponse_multiLayerStorage - Contains information about the storage destination.

$sel:httpStatus:PutStorageConfigurationResponse', putStorageConfigurationResponse_httpStatus - The response's http status code.

$sel:storageType:PutStorageConfiguration', putStorageConfigurationResponse_storageType - The type of storage that you specified for your data. The storage type can be one of the following values:

SITEWISE_DEFAULT_STORAGE – IoT SiteWise replicates your data into a service managed database.

MULTI_LAYER_STORAGE – IoT SiteWise replicates your data into a service managed database and saves a copy of your raw data and metadata in an Amazon S3 object that you specified.

$sel:configurationStatus:PutStorageConfigurationResponse', putStorageConfigurationResponse_configurationStatus - Undocumented member.

Response Lenses

putStorageConfigurationResponse_multiLayerStorage :: Lens' PutStorageConfigurationResponse (Maybe MultiLayerStorage)

Contains information about the storage destination.

putStorageConfigurationResponse_httpStatus :: Lens' PutStorageConfigurationResponse Int

The response's http status code.

putStorageConfigurationResponse_storageType :: Lens' PutStorageConfigurationResponse StorageType

The type of storage that you specified for your data. The storage type can be one of the following values:

SITEWISE_DEFAULT_STORAGE – IoT SiteWise replicates your data into a service managed database.

MULTI_LAYER_STORAGE – IoT SiteWise replicates your data into a service managed database and saves a copy of your raw data and metadata in an Amazon S3 object that you specified.

putStorageConfigurationResponse_configurationStatus :: Lens' PutStorageConfigurationResponse ConfigurationStatus

Undocumented member.