Network.AWS.IoTSiteWise.Types.Aggregates

Description

Documentation

data Aggregates

Contains the (pre-calculated) aggregate values for an asset property.

See: newAggregates smart constructor.

Constructors

Aggregates'

Fields

maximum :: Maybe Double

The maximum value of the time series over a time interval window.

average :: Maybe Double

The average (mean) value of the time series over a time interval window.

count :: Maybe Double

The count of data points in the time series over a time interval window.

minimum :: Maybe Double

The minimum value of the time series over a time interval window.

standardDeviation :: Maybe Double

The standard deviation of the time series over a time interval window.

sum :: Maybe Double

The sum of the time series over a time interval window.

newAggregates :: Aggregates

Create a value of Aggregates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maximum:Aggregates', aggregates_maximum - The maximum value of the time series over a time interval window.

$sel:average:Aggregates', aggregates_average - The average (mean) value of the time series over a time interval window.

$sel:count:Aggregates', aggregates_count - The count of data points in the time series over a time interval window.

$sel:minimum:Aggregates', aggregates_minimum - The minimum value of the time series over a time interval window.

$sel:standardDeviation:Aggregates', aggregates_standardDeviation - The standard deviation of the time series over a time interval window.

$sel:sum:Aggregates', aggregates_sum - The sum of the time series over a time interval window.

aggregates_maximum :: Lens' Aggregates (Maybe Double)

The maximum value of the time series over a time interval window.

aggregates_average :: Lens' Aggregates (Maybe Double)

The average (mean) value of the time series over a time interval window.

aggregates_count :: Lens' Aggregates (Maybe Double)

The count of data points in the time series over a time interval window.

aggregates_minimum :: Lens' Aggregates (Maybe Double)

The minimum value of the time series over a time interval window.

aggregates_standardDeviation :: Lens' Aggregates (Maybe Double)

The standard deviation of the time series over a time interval window.

aggregates_sum :: Lens' Aggregates (Maybe Double)

The sum of the time series over a time interval window.