Network.AWS.IoTSiteWise.Types.AssetProperty

Description

Documentation

data AssetProperty

Contains asset property information.

See: newAssetProperty smart constructor.

Constructors

AssetProperty'

Fields

dataTypeSpec :: Maybe Text

The data type of the structure for this property. This parameter exists on properties that have the STRUCT data type.

notification :: Maybe PropertyNotification

The asset property's notification topic and state. For more information, see UpdateAssetProperty.

alias :: Maybe Text

The alias that identifies the property, such as an OPC-UA server data stream path (for example, /company/windfarm/3/turbine/7/temperature). For more information, see Mapping industrial data streams to asset properties in the IoT SiteWise User Guide.

unit :: Maybe Text

The unit (such as Newtons or RPM) of the asset property.

id :: Text

The ID of the asset property.

name :: Text

The name of the property.

dataType :: PropertyDataType

The data type of the asset property.

newAssetProperty

Arguments

:: Text
$sel:id:AssetProperty'

-> Text
$sel:name:AssetProperty'

-> PropertyDataType
$sel:dataType:AssetProperty'

-> AssetProperty

Create a value of AssetProperty with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataTypeSpec:AssetProperty', assetProperty_dataTypeSpec - The data type of the structure for this property. This parameter exists on properties that have the STRUCT data type.

$sel:notification:AssetProperty', assetProperty_notification - The asset property's notification topic and state. For more information, see UpdateAssetProperty.

$sel:alias:AssetProperty', assetProperty_alias - The alias that identifies the property, such as an OPC-UA server data stream path (for example, /company/windfarm/3/turbine/7/temperature). For more information, see Mapping industrial data streams to asset properties in the IoT SiteWise User Guide.

$sel:unit:AssetProperty', assetProperty_unit - The unit (such as Newtons or RPM) of the asset property.

$sel:id:AssetProperty', assetProperty_id - The ID of the asset property.

$sel:name:AssetProperty', assetProperty_name - The name of the property.

$sel:dataType:AssetProperty', assetProperty_dataType - The data type of the asset property.

assetProperty_dataTypeSpec :: Lens' AssetProperty (Maybe Text)

The data type of the structure for this property. This parameter exists on properties that have the STRUCT data type.

assetProperty_notification :: Lens' AssetProperty (Maybe PropertyNotification)

The asset property's notification topic and state. For more information, see UpdateAssetProperty.

assetProperty_alias :: Lens' AssetProperty (Maybe Text)

The alias that identifies the property, such as an OPC-UA server data stream path (for example, /company/windfarm/3/turbine/7/temperature). For more information, see Mapping industrial data streams to asset properties in the IoT SiteWise User Guide.

assetProperty_unit :: Lens' AssetProperty (Maybe Text)

The unit (such as Newtons or RPM) of the asset property.

assetProperty_id :: Lens' AssetProperty Text

The ID of the asset property.

assetProperty_name :: Lens' AssetProperty Text

The name of the property.

assetProperty_dataType :: Lens' AssetProperty PropertyDataType

The data type of the asset property.