Network.AWS.IoTSiteWise.Types.AssetRelationshipSummary

Description

Documentation

data AssetRelationshipSummary

Contains information about assets that are related to one another.

See: newAssetRelationshipSummary smart constructor.

Constructors

AssetRelationshipSummary'

Fields

hierarchyInfo :: Maybe AssetHierarchyInfo

The assets that are related through an asset hierarchy.

This object is present if the relationshipType is HIERARCHY.

relationshipType :: AssetRelationshipType

The relationship type of the assets in this relationship. This value is one of the following:

HIERARCHY – The assets are related through an asset hierarchy. If you specify this relationship type, this asset relationship includes the hierarchyInfo object.

newAssetRelationshipSummary

Arguments

:: AssetRelationshipType
$sel:relationshipType:AssetRelationshipSummary'

-> AssetRelationshipSummary

Create a value of AssetRelationshipSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hierarchyInfo:AssetRelationshipSummary', assetRelationshipSummary_hierarchyInfo - The assets that are related through an asset hierarchy.

This object is present if the relationshipType is HIERARCHY.

$sel:relationshipType:AssetRelationshipSummary', assetRelationshipSummary_relationshipType - The relationship type of the assets in this relationship. This value is one of the following:

HIERARCHY – The assets are related through an asset hierarchy. If you specify this relationship type, this asset relationship includes the hierarchyInfo object.

assetRelationshipSummary_hierarchyInfo :: Lens' AssetRelationshipSummary (Maybe AssetHierarchyInfo)

The assets that are related through an asset hierarchy.

This object is present if the relationshipType is HIERARCHY.

assetRelationshipSummary_relationshipType :: Lens' AssetRelationshipSummary AssetRelationshipType

The relationship type of the assets in this relationship. This value is one of the following:

HIERARCHY – The assets are related through an asset hierarchy. If you specify this relationship type, this asset relationship includes the hierarchyInfo object.