Network.AWS.IoTSiteWise.Types.GatewaySummary

Description

Documentation

data GatewaySummary

Contains a summary of a gateway.

See: newGatewaySummary smart constructor.

Constructors

GatewaySummary'

Fields

gatewayPlatform :: Maybe GatewayPlatform

gatewayCapabilitySummaries :: Maybe [GatewayCapabilitySummary]

A list of gateway capability summaries that each contain a namespace and status. Each gateway capability defines data sources for the gateway. To retrieve a capability configuration's definition, use DescribeGatewayCapabilityConfiguration.

gatewayId :: Text

The ID of the gateway device.

gatewayName :: Text

The name of the asset.

creationDate :: POSIX

The date the gateway was created, in Unix epoch time.

lastUpdateDate :: POSIX

The date the gateway was last updated, in Unix epoch time.

newGatewaySummary

Arguments

:: Text
$sel:gatewayId:GatewaySummary'

-> Text
$sel:gatewayName:GatewaySummary'

-> UTCTime
$sel:creationDate:GatewaySummary'

-> UTCTime
$sel:lastUpdateDate:GatewaySummary'

-> GatewaySummary

Create a value of GatewaySummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:gatewayPlatform:GatewaySummary', gatewaySummary_gatewayPlatform - Undocumented member.

$sel:gatewayCapabilitySummaries:GatewaySummary', gatewaySummary_gatewayCapabilitySummaries - A list of gateway capability summaries that each contain a namespace and status. Each gateway capability defines data sources for the gateway. To retrieve a capability configuration's definition, use DescribeGatewayCapabilityConfiguration.

$sel:gatewayId:GatewaySummary', gatewaySummary_gatewayId - The ID of the gateway device.

$sel:gatewayName:GatewaySummary', gatewaySummary_gatewayName - The name of the asset.

$sel:creationDate:GatewaySummary', gatewaySummary_creationDate - The date the gateway was created, in Unix epoch time.

$sel:lastUpdateDate:GatewaySummary', gatewaySummary_lastUpdateDate - The date the gateway was last updated, in Unix epoch time.

gatewaySummary_gatewayPlatform :: Lens' GatewaySummary (Maybe GatewayPlatform)

Undocumented member.

gatewaySummary_gatewayCapabilitySummaries :: Lens' GatewaySummary (Maybe [GatewayCapabilitySummary])

A list of gateway capability summaries that each contain a namespace and status. Each gateway capability defines data sources for the gateway. To retrieve a capability configuration's definition, use DescribeGatewayCapabilityConfiguration.

gatewaySummary_gatewayId :: Lens' GatewaySummary Text

The ID of the gateway device.

gatewaySummary_gatewayName :: Lens' GatewaySummary Text

The name of the asset.

gatewaySummary_creationDate :: Lens' GatewaySummary UTCTime

The date the gateway was created, in Unix epoch time.

gatewaySummary_lastUpdateDate :: Lens' GatewaySummary UTCTime

The date the gateway was last updated, in Unix epoch time.