Network.AWS.IoTSiteWise.Types.PortalStatus

Description

Documentation

data PortalStatus

Contains information about the current status of a portal.

See: newPortalStatus smart constructor.

Constructors

PortalStatus'

Fields

error :: Maybe MonitorErrorDetails

Contains associated error information, if any.

state :: PortalState

The current state of the portal.

newPortalStatus

Arguments

:: PortalState
$sel:state:PortalStatus'

-> PortalStatus

Create a value of PortalStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:error:PortalStatus', portalStatus_error - Contains associated error information, if any.

$sel:state:PortalStatus', portalStatus_state - The current state of the portal.

portalStatus_error :: Lens' PortalStatus (Maybe MonitorErrorDetails)

Contains associated error information, if any.

portalStatus_state :: Lens' PortalStatus PortalState

The current state of the portal.