Network.AWS.IoTSiteWise.Types.PortalSummary

Description

Documentation

data PortalSummary

Contains a portal summary.

See: newPortalSummary smart constructor.

Constructors

PortalSummary'

Fields

lastUpdateDate :: Maybe POSIX

The date the portal was last updated, in Unix epoch time.

creationDate :: Maybe POSIX

The date the portal was created, in Unix epoch time.

description :: Maybe Text

The portal's description.

roleArn :: Maybe Text

The ARN of the service role that allows the portal's users to access your IoT SiteWise resources on your behalf. For more information, see Using service roles for IoT SiteWise Monitor in the IoT SiteWise User Guide.

id :: Text

The ID of the portal.

name :: Text

The name of the portal.

startUrl :: Text

The URL for the IoT SiteWise Monitor portal. You can use this URL to access portals that use Amazon Web Services SSO for authentication. For portals that use IAM for authentication, you must use the IoT SiteWise console to get a URL that you can use to access the portal.

status :: PortalStatus

newPortalSummary

Arguments

:: Text
$sel:id:PortalSummary'

-> Text
$sel:name:PortalSummary'

-> Text
$sel:startUrl:PortalSummary'

-> PortalStatus
$sel:status:PortalSummary'

-> PortalSummary

Create a value of PortalSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdateDate:PortalSummary', portalSummary_lastUpdateDate - The date the portal was last updated, in Unix epoch time.

$sel:creationDate:PortalSummary', portalSummary_creationDate - The date the portal was created, in Unix epoch time.

$sel:description:PortalSummary', portalSummary_description - The portal's description.

$sel:roleArn:PortalSummary', portalSummary_roleArn - The ARN of the service role that allows the portal's users to access your IoT SiteWise resources on your behalf. For more information, see Using service roles for IoT SiteWise Monitor in the IoT SiteWise User Guide.

$sel:id:PortalSummary', portalSummary_id - The ID of the portal.

$sel:name:PortalSummary', portalSummary_name - The name of the portal.

$sel:startUrl:PortalSummary', portalSummary_startUrl - The URL for the IoT SiteWise Monitor portal. You can use this URL to access portals that use Amazon Web Services SSO for authentication. For portals that use IAM for authentication, you must use the IoT SiteWise console to get a URL that you can use to access the portal.

$sel:status:PortalSummary', portalSummary_status - Undocumented member.

portalSummary_lastUpdateDate :: Lens' PortalSummary (Maybe UTCTime)

The date the portal was last updated, in Unix epoch time.

portalSummary_creationDate :: Lens' PortalSummary (Maybe UTCTime)

The date the portal was created, in Unix epoch time.

portalSummary_description :: Lens' PortalSummary (Maybe Text)

The portal's description.

portalSummary_roleArn :: Lens' PortalSummary (Maybe Text)

The ARN of the service role that allows the portal's users to access your IoT SiteWise resources on your behalf. For more information, see Using service roles for IoT SiteWise Monitor in the IoT SiteWise User Guide.

portalSummary_id :: Lens' PortalSummary Text

The ID of the portal.

portalSummary_name :: Lens' PortalSummary Text

The name of the portal.

portalSummary_startUrl :: Lens' PortalSummary Text

The URL for the IoT SiteWise Monitor portal. You can use this URL to access portals that use Amazon Web Services SSO for authentication. For portals that use IAM for authentication, you must use the IoT SiteWise console to get a URL that you can use to access the portal.

portalSummary_status :: Lens' PortalSummary PortalStatus

Undocumented member.