Network.AWS.IoTSiteWise.Types.PropertyType

Description

Documentation

data PropertyType

Contains a property type, which can be one of attribute, measurement, metric, or transform.

See: newPropertyType smart constructor.

Constructors

PropertyType'

Fields

attribute :: Maybe Attribute

Specifies an asset attribute property. An attribute generally contains static information, such as the serial number of an IIoT wind turbine.

transform :: Maybe Transform

Specifies an asset transform property. A transform contains a mathematical expression that maps a property's data points from one form to another, such as a unit conversion from Celsius to Fahrenheit.

metric :: Maybe Metric

Specifies an asset metric property. A metric contains a mathematical expression that uses aggregate functions to process all input data points over a time interval and output a single data point, such as to calculate the average hourly temperature.

measurement :: Maybe Measurement

Specifies an asset measurement property. A measurement represents a device's raw sensor data stream, such as timestamped temperature values or timestamped power values.

newPropertyType :: PropertyType

Create a value of PropertyType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attribute:PropertyType', propertyType_attribute - Specifies an asset attribute property. An attribute generally contains static information, such as the serial number of an IIoT wind turbine.

$sel:transform:PropertyType', propertyType_transform - Specifies an asset transform property. A transform contains a mathematical expression that maps a property's data points from one form to another, such as a unit conversion from Celsius to Fahrenheit.

$sel:metric:PropertyType', propertyType_metric - Specifies an asset metric property. A metric contains a mathematical expression that uses aggregate functions to process all input data points over a time interval and output a single data point, such as to calculate the average hourly temperature.

$sel:measurement:PropertyType', propertyType_measurement - Specifies an asset measurement property. A measurement represents a device's raw sensor data stream, such as timestamped temperature values or timestamped power values.

propertyType_attribute :: Lens' PropertyType (Maybe Attribute)

Specifies an asset attribute property. An attribute generally contains static information, such as the serial number of an IIoT wind turbine.

propertyType_transform :: Lens' PropertyType (Maybe Transform)

Specifies an asset transform property. A transform contains a mathematical expression that maps a property's data points from one form to another, such as a unit conversion from Celsius to Fahrenheit.

propertyType_metric :: Lens' PropertyType (Maybe Metric)

Specifies an asset metric property. A metric contains a mathematical expression that uses aggregate functions to process all input data points over a time interval and output a single data point, such as to calculate the average hourly temperature.

propertyType_measurement :: Lens' PropertyType (Maybe Measurement)

Specifies an asset measurement property. A measurement represents a device's raw sensor data stream, such as timestamped temperature values or timestamped power values.