Network.AWS.IoTSiteWise.UpdateAccessPolicy

Description

Updates an existing access policy that specifies an identity's access to an IoT SiteWise Monitor portal or project resource.

Creating a Request

data UpdateAccessPolicy

See: newUpdateAccessPolicy smart constructor.

Constructors

UpdateAccessPolicy'

Fields

clientToken :: Maybe Text

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

accessPolicyId :: Text

The ID of the access policy.

accessPolicyIdentity :: Identity

The identity for this access policy. Choose an Amazon Web Services SSO user, an Amazon Web Services SSO group, or an IAM user.

accessPolicyResource :: Resource

The IoT SiteWise Monitor resource for this access policy. Choose either a portal or a project.

accessPolicyPermission :: Permission

The permission level for this access policy. Note that a project ADMINISTRATOR is also known as a project owner.

newUpdateAccessPolicy

Arguments

:: Text
$sel:accessPolicyId:UpdateAccessPolicy'

-> Identity
$sel:accessPolicyIdentity:UpdateAccessPolicy'

-> Resource
$sel:accessPolicyResource:UpdateAccessPolicy'

-> Permission
$sel:accessPolicyPermission:UpdateAccessPolicy'

-> UpdateAccessPolicy

Create a value of UpdateAccessPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:UpdateAccessPolicy', updateAccessPolicy_clientToken - A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

$sel:accessPolicyId:UpdateAccessPolicy', updateAccessPolicy_accessPolicyId - The ID of the access policy.

$sel:accessPolicyIdentity:UpdateAccessPolicy', updateAccessPolicy_accessPolicyIdentity - The identity for this access policy. Choose an Amazon Web Services SSO user, an Amazon Web Services SSO group, or an IAM user.

$sel:accessPolicyResource:UpdateAccessPolicy', updateAccessPolicy_accessPolicyResource - The IoT SiteWise Monitor resource for this access policy. Choose either a portal or a project.

$sel:accessPolicyPermission:UpdateAccessPolicy', updateAccessPolicy_accessPolicyPermission - The permission level for this access policy. Note that a project ADMINISTRATOR is also known as a project owner.

Request Lenses

updateAccessPolicy_clientToken :: Lens' UpdateAccessPolicy (Maybe Text)

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

updateAccessPolicy_accessPolicyId :: Lens' UpdateAccessPolicy Text

The ID of the access policy.

updateAccessPolicy_accessPolicyIdentity :: Lens' UpdateAccessPolicy Identity

The identity for this access policy. Choose an Amazon Web Services SSO user, an Amazon Web Services SSO group, or an IAM user.

updateAccessPolicy_accessPolicyResource :: Lens' UpdateAccessPolicy Resource

The IoT SiteWise Monitor resource for this access policy. Choose either a portal or a project.

updateAccessPolicy_accessPolicyPermission :: Lens' UpdateAccessPolicy Permission

The permission level for this access policy. Note that a project ADMINISTRATOR is also known as a project owner.

Destructuring the Response

data UpdateAccessPolicyResponse

See: newUpdateAccessPolicyResponse smart constructor.

Constructors

UpdateAccessPolicyResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateAccessPolicyResponse

Arguments

:: Int
$sel:httpStatus:UpdateAccessPolicyResponse'

-> UpdateAccessPolicyResponse

Create a value of UpdateAccessPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateAccessPolicyResponse', updateAccessPolicyResponse_httpStatus - The response's http status code.

Response Lenses

updateAccessPolicyResponse_httpStatus :: Lens' UpdateAccessPolicyResponse Int

The response's http status code.