Network.AWS.IoTSiteWise.UpdateAssetProperty

Description

Updates an asset property's alias and notification state.

This operation overwrites the property's existing alias and notification state. To keep your existing property's alias or notification state, you must include the existing values in the UpdateAssetProperty request. For more information, see DescribeAssetProperty.

Creating a Request

data UpdateAssetProperty

See: newUpdateAssetProperty smart constructor.

Constructors

UpdateAssetProperty'

Fields

clientToken :: Maybe Text

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

propertyNotificationState :: Maybe PropertyNotificationState

The MQTT notification state (enabled or disabled) for this asset property. When the notification state is enabled, IoT SiteWise publishes property value updates to a unique MQTT topic. For more information, see Interacting with other services in the IoT SiteWise User Guide.

If you omit this parameter, the notification state is set to DISABLED.

propertyAlias :: Maybe Text

The alias that identifies the property, such as an OPC-UA server data stream path (for example, /company/windfarm/3/turbine/7/temperature). For more information, see Mapping industrial data streams to asset properties in the IoT SiteWise User Guide.

If you omit this parameter, the alias is removed from the property.

assetId :: Text

The ID of the asset to be updated.

propertyId :: Text

The ID of the asset property to be updated.

newUpdateAssetProperty

Arguments

:: Text
$sel:assetId:UpdateAssetProperty'

-> Text
$sel:propertyId:UpdateAssetProperty'

-> UpdateAssetProperty

Create a value of UpdateAssetProperty with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:UpdateAssetProperty', updateAssetProperty_clientToken - A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

$sel:propertyNotificationState:UpdateAssetProperty', updateAssetProperty_propertyNotificationState - The MQTT notification state (enabled or disabled) for this asset property. When the notification state is enabled, IoT SiteWise publishes property value updates to a unique MQTT topic. For more information, see Interacting with other services in the IoT SiteWise User Guide.

If you omit this parameter, the notification state is set to DISABLED.

$sel:propertyAlias:UpdateAssetProperty', updateAssetProperty_propertyAlias - The alias that identifies the property, such as an OPC-UA server data stream path (for example, /company/windfarm/3/turbine/7/temperature). For more information, see Mapping industrial data streams to asset properties in the IoT SiteWise User Guide.

If you omit this parameter, the alias is removed from the property.

$sel:assetId:UpdateAssetProperty', updateAssetProperty_assetId - The ID of the asset to be updated.

$sel:propertyId:UpdateAssetProperty', updateAssetProperty_propertyId - The ID of the asset property to be updated.

Request Lenses

updateAssetProperty_clientToken :: Lens' UpdateAssetProperty (Maybe Text)

A unique case-sensitive identifier that you can provide to ensure the idempotency of the request. Don't reuse this client token if a new idempotent request is required.

updateAssetProperty_propertyNotificationState :: Lens' UpdateAssetProperty (Maybe PropertyNotificationState)

The MQTT notification state (enabled or disabled) for this asset property. When the notification state is enabled, IoT SiteWise publishes property value updates to a unique MQTT topic. For more information, see Interacting with other services in the IoT SiteWise User Guide.

If you omit this parameter, the notification state is set to DISABLED.

updateAssetProperty_propertyAlias :: Lens' UpdateAssetProperty (Maybe Text)

The alias that identifies the property, such as an OPC-UA server data stream path (for example, /company/windfarm/3/turbine/7/temperature). For more information, see Mapping industrial data streams to asset properties in the IoT SiteWise User Guide.

If you omit this parameter, the alias is removed from the property.

updateAssetProperty_assetId :: Lens' UpdateAssetProperty Text

The ID of the asset to be updated.

updateAssetProperty_propertyId :: Lens' UpdateAssetProperty Text

The ID of the asset property to be updated.

Destructuring the Response

data UpdateAssetPropertyResponse

See: newUpdateAssetPropertyResponse smart constructor.

Constructors

UpdateAssetPropertyResponse'

newUpdateAssetPropertyResponse :: UpdateAssetPropertyResponse

Create a value of UpdateAssetPropertyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.