Network.AWS.IoTThingsGraph.CreateFlowTemplate

Description

Creates a workflow template. Workflows can be created only in the user's namespace. (The public namespace contains only entities.) The workflow can contain only entities in the specified namespace. The workflow is validated against the entities in the latest version of the user's namespace unless another namespace version is specified in the request.

Creating a Request

data CreateFlowTemplate

See: newCreateFlowTemplate smart constructor.

Constructors

CreateFlowTemplate'

Fields

compatibleNamespaceVersion :: Maybe Integer

The namespace version in which the workflow is to be created.

If no value is specified, the latest version is used by default.

definition :: DefinitionDocument

The workflow DefinitionDocument.

newCreateFlowTemplate

Arguments

:: DefinitionDocument
$sel:definition:CreateFlowTemplate'

-> CreateFlowTemplate

Create a value of CreateFlowTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:compatibleNamespaceVersion:CreateFlowTemplate', createFlowTemplate_compatibleNamespaceVersion - The namespace version in which the workflow is to be created.

If no value is specified, the latest version is used by default.

$sel:definition:CreateFlowTemplate', createFlowTemplate_definition - The workflow DefinitionDocument.

Request Lenses

createFlowTemplate_compatibleNamespaceVersion :: Lens' CreateFlowTemplate (Maybe Integer)

The namespace version in which the workflow is to be created.

If no value is specified, the latest version is used by default.

createFlowTemplate_definition :: Lens' CreateFlowTemplate DefinitionDocument

The workflow DefinitionDocument.

Destructuring the Response

data CreateFlowTemplateResponse

See: newCreateFlowTemplateResponse smart constructor.

Constructors

CreateFlowTemplateResponse'

Fields

summary :: Maybe FlowTemplateSummary

The summary object that describes the created workflow.

httpStatus :: Int

The response's http status code.

newCreateFlowTemplateResponse

Arguments

:: Int
$sel:httpStatus:CreateFlowTemplateResponse'

-> CreateFlowTemplateResponse

Create a value of CreateFlowTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:summary:CreateFlowTemplateResponse', createFlowTemplateResponse_summary - The summary object that describes the created workflow.

$sel:httpStatus:CreateFlowTemplateResponse', createFlowTemplateResponse_httpStatus - The response's http status code.

Response Lenses

createFlowTemplateResponse_summary :: Lens' CreateFlowTemplateResponse (Maybe FlowTemplateSummary)

The summary object that describes the created workflow.

createFlowTemplateResponse_httpStatus :: Lens' CreateFlowTemplateResponse Int

The response's http status code.