Network.AWS.IoTThingsGraph.Lens

Description

Operations

GetFlowTemplate

getFlowTemplate_revisionNumber :: Lens' GetFlowTemplate (Maybe Integer)

The number of the workflow revision to retrieve.

getFlowTemplate_id :: Lens' GetFlowTemplate Text

The ID of the workflow.

The ID should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:workflow:WORKFLOWNAME

getFlowTemplateResponse_description :: Lens' GetFlowTemplateResponse (Maybe FlowTemplateDescription)

The object that describes the specified workflow.

getFlowTemplateResponse_httpStatus :: Lens' GetFlowTemplateResponse Int

The response's http status code.

UpdateSystemTemplate

updateSystemTemplate_compatibleNamespaceVersion :: Lens' UpdateSystemTemplate (Maybe Integer)

The version of the user's namespace. Defaults to the latest version of the user's namespace.

If no value is specified, the latest version is used by default.

updateSystemTemplate_id :: Lens' UpdateSystemTemplate Text

The ID of the system to be updated.

The ID should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:system:SYSTEMNAME

updateSystemTemplate_definition :: Lens' UpdateSystemTemplate DefinitionDocument

The DefinitionDocument that contains the updated system definition.

updateSystemTemplateResponse_summary :: Lens' UpdateSystemTemplateResponse (Maybe SystemTemplateSummary)

An object containing summary information about the updated system.

updateSystemTemplateResponse_httpStatus :: Lens' UpdateSystemTemplateResponse Int

The response's http status code.

DeleteSystemTemplate

deleteSystemTemplate_id :: Lens' DeleteSystemTemplate Text

The ID of the system to be deleted.

The ID should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:system:SYSTEMNAME

deleteSystemTemplateResponse_httpStatus :: Lens' DeleteSystemTemplateResponse Int

The response's http status code.

DeprecateFlowTemplate

deprecateFlowTemplate_id :: Lens' DeprecateFlowTemplate Text

The ID of the workflow to be deleted.

The ID should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:workflow:WORKFLOWNAME

deprecateFlowTemplateResponse_httpStatus :: Lens' DeprecateFlowTemplateResponse Int

The response's http status code.

DeploySystemInstance

deploySystemInstance_id :: Lens' DeploySystemInstance (Maybe Text)

The ID of the system instance. This value is returned by the CreateSystemInstance action.

The ID should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:deployment:DEPLOYMENTNAME

deploySystemInstanceResponse_greengrassDeploymentId :: Lens' DeploySystemInstanceResponse (Maybe Text)

The ID of the Greengrass deployment used to deploy the system instance.

deploySystemInstanceResponse_httpStatus :: Lens' DeploySystemInstanceResponse Int

The response's http status code.

deploySystemInstanceResponse_summary :: Lens' DeploySystemInstanceResponse SystemInstanceSummary

An object that contains summary information about a system instance that was deployed.

SearchFlowTemplates

searchFlowTemplates_filters :: Lens' SearchFlowTemplates (Maybe [FlowTemplateFilter])

An array of objects that limit the result set. The only valid filter is DEVICE_MODEL_ID.

searchFlowTemplates_nextToken :: Lens' SearchFlowTemplates (Maybe Text)

The string that specifies the next page of results. Use this when you're paginating results.

searchFlowTemplates_maxResults :: Lens' SearchFlowTemplates (Maybe Natural)

The maximum number of results to return in the response.

searchFlowTemplatesResponse_nextToken :: Lens' SearchFlowTemplatesResponse (Maybe Text)

The string to specify as nextToken when you request the next page of results.

searchFlowTemplatesResponse_summaries :: Lens' SearchFlowTemplatesResponse (Maybe [FlowTemplateSummary])

An array of objects that contain summary information about each workflow in the result set.

searchFlowTemplatesResponse_httpStatus :: Lens' SearchFlowTemplatesResponse Int

The response's http status code.

DeleteNamespace

deleteNamespaceResponse_namespaceArn :: Lens' DeleteNamespaceResponse (Maybe Text)

The ARN of the namespace to be deleted.

deleteNamespaceResponse_namespaceName :: Lens' DeleteNamespaceResponse (Maybe Text)

The name of the namespace to be deleted.

deleteNamespaceResponse_httpStatus :: Lens' DeleteNamespaceResponse Int

The response's http status code.

GetSystemInstance

getSystemInstance_id :: Lens' GetSystemInstance Text

The ID of the system deployment instance. This value is returned by CreateSystemInstance.

The ID should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:deployment:DEPLOYMENTNAME

getSystemInstanceResponse_description :: Lens' GetSystemInstanceResponse (Maybe SystemInstanceDescription)

An object that describes the system instance.

getSystemInstanceResponse_httpStatus :: Lens' GetSystemInstanceResponse Int

The response's http status code.

ListFlowExecutionMessages

listFlowExecutionMessages_nextToken :: Lens' ListFlowExecutionMessages (Maybe Text)

The string that specifies the next page of results. Use this when you're paginating results.

listFlowExecutionMessages_maxResults :: Lens' ListFlowExecutionMessages (Maybe Natural)

The maximum number of results to return in the response.

listFlowExecutionMessages_flowExecutionId :: Lens' ListFlowExecutionMessages Text

The ID of the flow execution.

listFlowExecutionMessagesResponse_nextToken :: Lens' ListFlowExecutionMessagesResponse (Maybe Text)

The string to specify as nextToken when you request the next page of results.

listFlowExecutionMessagesResponse_messages :: Lens' ListFlowExecutionMessagesResponse (Maybe [FlowExecutionMessage])

A list of objects that contain information about events in the specified flow execution.

listFlowExecutionMessagesResponse_httpStatus :: Lens' ListFlowExecutionMessagesResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_nextToken :: Lens' ListTagsForResource (Maybe Text)

The token that specifies the next page of results to return.

listTagsForResource_maxResults :: Lens' ListTagsForResource (Maybe Natural)

The maximum number of tags to return.

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the resource whose tags are to be returned.

listTagsForResourceResponse_nextToken :: Lens' ListTagsForResourceResponse (Maybe Text)

The token that specifies the next page of results to return.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe [Tag])

List of tags returned by the ListTagsForResource operation.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

SearchFlowExecutions

searchFlowExecutions_startTime :: Lens' SearchFlowExecutions (Maybe UTCTime)

The date and time of the earliest flow execution to return.

searchFlowExecutions_flowExecutionId :: Lens' SearchFlowExecutions (Maybe Text)

The ID of a flow execution.

searchFlowExecutions_nextToken :: Lens' SearchFlowExecutions (Maybe Text)

The string that specifies the next page of results. Use this when you're paginating results.

searchFlowExecutions_endTime :: Lens' SearchFlowExecutions (Maybe UTCTime)

The date and time of the latest flow execution to return.

searchFlowExecutions_maxResults :: Lens' SearchFlowExecutions (Maybe Natural)

The maximum number of results to return in the response.

searchFlowExecutions_systemInstanceId :: Lens' SearchFlowExecutions Text

The ID of the system instance that contains the flow.

searchFlowExecutionsResponse_nextToken :: Lens' SearchFlowExecutionsResponse (Maybe Text)

The string to specify as nextToken when you request the next page of results.

searchFlowExecutionsResponse_summaries :: Lens' SearchFlowExecutionsResponse (Maybe [FlowExecutionSummary])

An array of objects that contain summary information about each workflow execution in the result set.

searchFlowExecutionsResponse_httpStatus :: Lens' SearchFlowExecutionsResponse Int

The response's http status code.

DeleteSystemInstance

deleteSystemInstance_id :: Lens' DeleteSystemInstance (Maybe Text)

The ID of the system instance to be deleted.

deleteSystemInstanceResponse_httpStatus :: Lens' DeleteSystemInstanceResponse Int

The response's http status code.

CreateSystemInstance

createSystemInstance_metricsConfiguration :: Lens' CreateSystemInstance (Maybe MetricsConfiguration)

Undocumented member.

createSystemInstance_greengrassGroupName :: Lens' CreateSystemInstance (Maybe Text)

The name of the Greengrass group where the system instance will be deployed. This value is required if the value of the target parameter is GREENGRASS.

createSystemInstance_flowActionsRoleArn :: Lens' CreateSystemInstance (Maybe Text)

The ARN of the IAM role that AWS IoT Things Graph will assume when it executes the flow. This role must have read and write access to AWS Lambda and AWS IoT and any other AWS services that the flow uses when it executes. This value is required if the value of the target parameter is CLOUD.

createSystemInstance_s3BucketName :: Lens' CreateSystemInstance (Maybe Text)

The name of the Amazon Simple Storage Service bucket that will be used to store and deploy the system instance's resource file. This value is required if the value of the target parameter is GREENGRASS.

createSystemInstance_tags :: Lens' CreateSystemInstance (Maybe [Tag])

Metadata, consisting of key-value pairs, that can be used to categorize your system instances.

createSystemInstance_definition :: Lens' CreateSystemInstance DefinitionDocument

Undocumented member.

createSystemInstance_target :: Lens' CreateSystemInstance DeploymentTarget

The target type of the deployment. Valid values are GREENGRASS and CLOUD.

createSystemInstanceResponse_summary :: Lens' CreateSystemInstanceResponse (Maybe SystemInstanceSummary)

The summary object that describes the new system instance.

createSystemInstanceResponse_httpStatus :: Lens' CreateSystemInstanceResponse Int

The response's http status code.

DeprecateSystemTemplate

deprecateSystemTemplate_id :: Lens' DeprecateSystemTemplate Text

The ID of the system to delete.

The ID should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:system:SYSTEMNAME

deprecateSystemTemplateResponse_httpStatus :: Lens' DeprecateSystemTemplateResponse Int

The response's http status code.

GetSystemTemplateRevisions

getSystemTemplateRevisions_nextToken :: Lens' GetSystemTemplateRevisions (Maybe Text)

The string that specifies the next page of results. Use this when you're paginating results.

getSystemTemplateRevisions_maxResults :: Lens' GetSystemTemplateRevisions (Maybe Natural)

The maximum number of results to return in the response.

getSystemTemplateRevisions_id :: Lens' GetSystemTemplateRevisions Text

The ID of the system template.

The ID should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:system:SYSTEMNAME

getSystemTemplateRevisionsResponse_nextToken :: Lens' GetSystemTemplateRevisionsResponse (Maybe Text)

The string to specify as nextToken when you request the next page of results.

getSystemTemplateRevisionsResponse_summaries :: Lens' GetSystemTemplateRevisionsResponse (Maybe [SystemTemplateSummary])

An array of objects that contain summary data about the system template revisions.

getSystemTemplateRevisionsResponse_httpStatus :: Lens' GetSystemTemplateRevisionsResponse Int

The response's http status code.

SearchEntities

searchEntities_filters :: Lens' SearchEntities (Maybe [EntityFilter])

Optional filter to apply to the search. Valid filters are NAME NAMESPACE, SEMANTIC_TYPE_PATH and REFERENCED_ENTITY_ID. REFERENCED_ENTITY_ID filters on entities that are used by the entity in the result set. For example, you can filter on the ID of a property that is used in a state.

Multiple filters function as OR criteria in the query. Multiple values passed inside the filter function as AND criteria.

searchEntities_namespaceVersion :: Lens' SearchEntities (Maybe Integer)

The version of the user's namespace. Defaults to the latest version of the user's namespace.

searchEntities_nextToken :: Lens' SearchEntities (Maybe Text)

The string that specifies the next page of results. Use this when you're paginating results.

searchEntities_maxResults :: Lens' SearchEntities (Maybe Natural)

The maximum number of results to return in the response.

searchEntities_entityTypes :: Lens' SearchEntities [EntityType]

The entity types for which to search.

searchEntitiesResponse_nextToken :: Lens' SearchEntitiesResponse (Maybe Text)

The string to specify as nextToken when you request the next page of results.

searchEntitiesResponse_descriptions :: Lens' SearchEntitiesResponse (Maybe [EntityDescription])

An array of descriptions for each entity returned in the search result.

searchEntitiesResponse_httpStatus :: Lens' SearchEntitiesResponse Int

The response's http status code.

DeleteFlowTemplate

deleteFlowTemplate_id :: Lens' DeleteFlowTemplate Text

The ID of the workflow to be deleted.

The ID should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:workflow:WORKFLOWNAME

deleteFlowTemplateResponse_httpStatus :: Lens' DeleteFlowTemplateResponse Int

The response's http status code.

UpdateFlowTemplate

updateFlowTemplate_compatibleNamespaceVersion :: Lens' UpdateFlowTemplate (Maybe Integer)

The version of the user's namespace.

If no value is specified, the latest version is used by default. Use the GetFlowTemplateRevisions if you want to find earlier revisions of the flow to update.

updateFlowTemplate_id :: Lens' UpdateFlowTemplate Text

The ID of the workflow to be updated.

The ID should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:workflow:WORKFLOWNAME

updateFlowTemplate_definition :: Lens' UpdateFlowTemplate DefinitionDocument

The DefinitionDocument that contains the updated workflow definition.

updateFlowTemplateResponse_summary :: Lens' UpdateFlowTemplateResponse (Maybe FlowTemplateSummary)

An object containing summary information about the updated workflow.

updateFlowTemplateResponse_httpStatus :: Lens' UpdateFlowTemplateResponse Int

The response's http status code.

GetSystemTemplate

getSystemTemplate_revisionNumber :: Lens' GetSystemTemplate (Maybe Integer)

The number that specifies the revision of the system to get.

getSystemTemplate_id :: Lens' GetSystemTemplate Text

The ID of the system to get. This ID must be in the user's namespace.

The ID should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:system:SYSTEMNAME

getSystemTemplateResponse_description :: Lens' GetSystemTemplateResponse (Maybe SystemTemplateDescription)

An object that contains summary data about the system.

getSystemTemplateResponse_httpStatus :: Lens' GetSystemTemplateResponse Int

The response's http status code.

SearchSystemInstances

searchSystemInstances_filters :: Lens' SearchSystemInstances (Maybe [SystemInstanceFilter])

Optional filter to apply to the search. Valid filters are SYSTEM_TEMPLATE_ID, STATUS, and GREENGRASS_GROUP_NAME.

Multiple filters function as OR criteria in the query. Multiple values passed inside the filter function as AND criteria.

searchSystemInstances_nextToken :: Lens' SearchSystemInstances (Maybe Text)

The string that specifies the next page of results. Use this when you're paginating results.

searchSystemInstances_maxResults :: Lens' SearchSystemInstances (Maybe Natural)

The maximum number of results to return in the response.

searchSystemInstancesResponse_nextToken :: Lens' SearchSystemInstancesResponse (Maybe Text)

The string to specify as nextToken when you request the next page of results.

searchSystemInstancesResponse_summaries :: Lens' SearchSystemInstancesResponse (Maybe [SystemInstanceSummary])

An array of objects that contain summary data abour the system instances in the result set.

searchSystemInstancesResponse_httpStatus :: Lens' SearchSystemInstancesResponse Int

The response's http status code.

GetUploadStatus

getUploadStatus_uploadId :: Lens' GetUploadStatus Text

The ID of the upload. This value is returned by the UploadEntityDefinitions action.

getUploadStatusResponse_failureReason :: Lens' GetUploadStatusResponse (Maybe [Text])

The reason for an upload failure.

getUploadStatusResponse_namespaceArn :: Lens' GetUploadStatusResponse (Maybe Text)

The ARN of the upload.

getUploadStatusResponse_namespaceVersion :: Lens' GetUploadStatusResponse (Maybe Integer)

The version of the user's namespace. Defaults to the latest version of the user's namespace.

getUploadStatusResponse_namespaceName :: Lens' GetUploadStatusResponse (Maybe Text)

The name of the upload's namespace.

getUploadStatusResponse_httpStatus :: Lens' GetUploadStatusResponse Int

The response's http status code.

getUploadStatusResponse_uploadId :: Lens' GetUploadStatusResponse Text

The ID of the upload.

getUploadStatusResponse_uploadStatus :: Lens' GetUploadStatusResponse UploadStatus

The status of the upload. The initial status is IN_PROGRESS. The response show all validation failures if the upload fails.

getUploadStatusResponse_createdDate :: Lens' GetUploadStatusResponse UTCTime

The date at which the upload was created.

CreateSystemTemplate

createSystemTemplate_compatibleNamespaceVersion :: Lens' CreateSystemTemplate (Maybe Integer)

The namespace version in which the system is to be created.

If no value is specified, the latest version is used by default.

createSystemTemplate_definition :: Lens' CreateSystemTemplate DefinitionDocument

The DefinitionDocument used to create the system.

createSystemTemplateResponse_summary :: Lens' CreateSystemTemplateResponse (Maybe SystemTemplateSummary)

The summary object that describes the created system.

createSystemTemplateResponse_httpStatus :: Lens' CreateSystemTemplateResponse Int

The response's http status code.

UndeploySystemInstance

undeploySystemInstance_id :: Lens' UndeploySystemInstance (Maybe Text)

The ID of the system instance to remove from its target.

undeploySystemInstanceResponse_summary :: Lens' UndeploySystemInstanceResponse (Maybe SystemInstanceSummary)

An object that contains summary information about the system instance that was removed from its target.

undeploySystemInstanceResponse_httpStatus :: Lens' UndeploySystemInstanceResponse Int

The response's http status code.

GetFlowTemplateRevisions

getFlowTemplateRevisions_nextToken :: Lens' GetFlowTemplateRevisions (Maybe Text)

The string that specifies the next page of results. Use this when you're paginating results.

getFlowTemplateRevisions_maxResults :: Lens' GetFlowTemplateRevisions (Maybe Natural)

The maximum number of results to return in the response.

getFlowTemplateRevisions_id :: Lens' GetFlowTemplateRevisions Text

The ID of the workflow.

The ID should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:workflow:WORKFLOWNAME

getFlowTemplateRevisionsResponse_nextToken :: Lens' GetFlowTemplateRevisionsResponse (Maybe Text)

The string to specify as nextToken when you request the next page of results.

getFlowTemplateRevisionsResponse_summaries :: Lens' GetFlowTemplateRevisionsResponse (Maybe [FlowTemplateSummary])

An array of objects that provide summary data about each revision.

getFlowTemplateRevisionsResponse_httpStatus :: Lens' GetFlowTemplateRevisionsResponse Int

The response's http status code.

GetNamespaceDeletionStatus

getNamespaceDeletionStatusResponse_status :: Lens' GetNamespaceDeletionStatusResponse (Maybe NamespaceDeletionStatus)

The status of the deletion request.

getNamespaceDeletionStatusResponse_namespaceArn :: Lens' GetNamespaceDeletionStatusResponse (Maybe Text)

The ARN of the namespace that is being deleted.

getNamespaceDeletionStatusResponse_namespaceName :: Lens' GetNamespaceDeletionStatusResponse (Maybe Text)

The name of the namespace that is being deleted.

getNamespaceDeletionStatusResponse_errorCode :: Lens' GetNamespaceDeletionStatusResponse (Maybe NamespaceDeletionStatusErrorCodes)

An error code returned by the namespace deletion task.

getNamespaceDeletionStatusResponse_errorMessage :: Lens' GetNamespaceDeletionStatusResponse (Maybe Text)

An error code returned by the namespace deletion task.

getNamespaceDeletionStatusResponse_httpStatus :: Lens' GetNamespaceDeletionStatusResponse Int

The response's http status code.

AssociateEntityToThing

associateEntityToThing_namespaceVersion :: Lens' AssociateEntityToThing (Maybe Integer)

The version of the user's namespace. Defaults to the latest version of the user's namespace.

associateEntityToThing_thingName :: Lens' AssociateEntityToThing Text

The name of the thing to which the entity is to be associated.

associateEntityToThing_entityId :: Lens' AssociateEntityToThing Text

The ID of the device to be associated with the thing.

The ID should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:device:DEVICENAME

associateEntityToThingResponse_httpStatus :: Lens' AssociateEntityToThingResponse Int

The response's http status code.

SearchSystemTemplates

searchSystemTemplates_filters :: Lens' SearchSystemTemplates (Maybe [SystemTemplateFilter])

An array of filters that limit the result set. The only valid filter is FLOW_TEMPLATE_ID.

searchSystemTemplates_nextToken :: Lens' SearchSystemTemplates (Maybe Text)

The string that specifies the next page of results. Use this when you're paginating results.

searchSystemTemplates_maxResults :: Lens' SearchSystemTemplates (Maybe Natural)

The maximum number of results to return in the response.

searchSystemTemplatesResponse_nextToken :: Lens' SearchSystemTemplatesResponse (Maybe Text)

The string to specify as nextToken when you request the next page of results.

searchSystemTemplatesResponse_summaries :: Lens' SearchSystemTemplatesResponse (Maybe [SystemTemplateSummary])

An array of objects that contain summary information about each system deployment in the result set.

searchSystemTemplatesResponse_httpStatus :: Lens' SearchSystemTemplatesResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the resource whose tags are returned.

tagResource_tags :: Lens' TagResource [Tag]

A list of tags to add to the resource.>

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the resource whose tags are to be removed.

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

A list of tag key names to remove from the resource. You don't specify the value. Both the key and its associated value are removed.

This parameter to the API requires a JSON text string argument. For information on how to format a JSON parameter for the various command line tool environments, see Using JSON for Parameters in the AWS CLI User Guide.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

GetEntities

getEntities_namespaceVersion :: Lens' GetEntities (Maybe Integer)

The version of the user's namespace. Defaults to the latest version of the user's namespace.

getEntities_ids :: Lens' GetEntities [Text]

An array of entity IDs.

The IDs should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:device:DEVICENAME

getEntitiesResponse_descriptions :: Lens' GetEntitiesResponse (Maybe [EntityDescription])

An array of descriptions for the specified entities.

getEntitiesResponse_httpStatus :: Lens' GetEntitiesResponse Int

The response's http status code.

DescribeNamespace

describeNamespace_namespaceName :: Lens' DescribeNamespace (Maybe Text)

The name of the user's namespace. Set this to aws to get the public namespace.

describeNamespaceResponse_namespaceArn :: Lens' DescribeNamespaceResponse (Maybe Text)

The ARN of the namespace.

describeNamespaceResponse_trackingNamespaceVersion :: Lens' DescribeNamespaceResponse (Maybe Integer)

The version of the public namespace that the latest version is tracking.

describeNamespaceResponse_namespaceVersion :: Lens' DescribeNamespaceResponse (Maybe Integer)

The version of the user's namespace to describe.

describeNamespaceResponse_namespaceName :: Lens' DescribeNamespaceResponse (Maybe Text)

The name of the namespace.

describeNamespaceResponse_trackingNamespaceName :: Lens' DescribeNamespaceResponse (Maybe Text)

The name of the public namespace that the latest namespace version is tracking.

describeNamespaceResponse_httpStatus :: Lens' DescribeNamespaceResponse Int

The response's http status code.

CreateFlowTemplate

createFlowTemplate_compatibleNamespaceVersion :: Lens' CreateFlowTemplate (Maybe Integer)

The namespace version in which the workflow is to be created.

If no value is specified, the latest version is used by default.

createFlowTemplate_definition :: Lens' CreateFlowTemplate DefinitionDocument

The workflow DefinitionDocument.

createFlowTemplateResponse_summary :: Lens' CreateFlowTemplateResponse (Maybe FlowTemplateSummary)

The summary object that describes the created workflow.

createFlowTemplateResponse_httpStatus :: Lens' CreateFlowTemplateResponse Int

The response's http status code.

UploadEntityDefinitions

uploadEntityDefinitions_syncWithPublicNamespace :: Lens' UploadEntityDefinitions (Maybe Bool)

A Boolean that specifies whether to synchronize with the latest version of the public namespace. If set to true, the upload will create a new namespace version.

uploadEntityDefinitions_deprecateExistingEntities :: Lens' UploadEntityDefinitions (Maybe Bool)

A Boolean that specifies whether to deprecate all entities in the latest version before uploading the new DefinitionDocument. If set to true, the upload will create a new namespace version.

uploadEntityDefinitions_document :: Lens' UploadEntityDefinitions (Maybe DefinitionDocument)

The DefinitionDocument that defines the updated entities.

uploadEntityDefinitionsResponse_httpStatus :: Lens' UploadEntityDefinitionsResponse Int

The response's http status code.

uploadEntityDefinitionsResponse_uploadId :: Lens' UploadEntityDefinitionsResponse Text

The ID that specifies the upload action. You can use this to track the status of the upload.

DissociateEntityFromThing

dissociateEntityFromThing_thingName :: Lens' DissociateEntityFromThing Text

The name of the thing to disassociate.

dissociateEntityFromThing_entityType :: Lens' DissociateEntityFromThing EntityType

The entity type from which to disassociate the thing.

dissociateEntityFromThingResponse_httpStatus :: Lens' DissociateEntityFromThingResponse Int

The response's http status code.

SearchThings

searchThings_namespaceVersion :: Lens' SearchThings (Maybe Integer)

The version of the user's namespace. Defaults to the latest version of the user's namespace.

searchThings_nextToken :: Lens' SearchThings (Maybe Text)

The string that specifies the next page of results. Use this when you're paginating results.

searchThings_maxResults :: Lens' SearchThings (Maybe Natural)

The maximum number of results to return in the response.

searchThings_entityId :: Lens' SearchThings Text

The ID of the entity to which the things are associated.

The IDs should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:device:DEVICENAME

searchThingsResponse_nextToken :: Lens' SearchThingsResponse (Maybe Text)

The string to specify as nextToken when you request the next page of results.

searchThingsResponse_things :: Lens' SearchThingsResponse (Maybe [Thing])

An array of things in the result set.

searchThingsResponse_httpStatus :: Lens' SearchThingsResponse Int

The response's http status code.

Types

DefinitionDocument

definitionDocument_language :: Lens' DefinitionDocument DefinitionLanguage

The language used to define the entity. GRAPHQL is the only valid value.

definitionDocument_text :: Lens' DefinitionDocument Text

The GraphQL text that defines the entity.

DependencyRevision

dependencyRevision_revisionNumber :: Lens' DependencyRevision (Maybe Integer)

The revision number of the workflow or system.

dependencyRevision_id :: Lens' DependencyRevision (Maybe Text)

The ID of the workflow or system.

EntityDescription

entityDescription_arn :: Lens' EntityDescription (Maybe Text)

The entity ARN.

entityDescription_createdAt :: Lens' EntityDescription (Maybe UTCTime)

The time at which the entity was created.

entityDescription_definition :: Lens' EntityDescription (Maybe DefinitionDocument)

The definition document of the entity.

entityDescription_id :: Lens' EntityDescription (Maybe Text)

The entity ID.

entityDescription_type :: Lens' EntityDescription (Maybe EntityType)

The entity type.

EntityFilter

entityFilter_value :: Lens' EntityFilter (Maybe [Text])

An array of string values for the search filter field. Multiple values function as AND criteria in the search.

entityFilter_name :: Lens' EntityFilter (Maybe EntityFilterName)

The name of the entity search filter field. REFERENCED_ENTITY_ID filters on entities that are used by the entity in the result set. For example, you can filter on the ID of a property that is used in a state.

FlowExecutionMessage

flowExecutionMessage_payload :: Lens' FlowExecutionMessage (Maybe Text)

A string containing information about the flow event.

flowExecutionMessage_eventType :: Lens' FlowExecutionMessage (Maybe FlowExecutionEventType)

The type of flow event .

flowExecutionMessage_timestamp :: Lens' FlowExecutionMessage (Maybe UTCTime)

The date and time when the message was last updated.

flowExecutionMessage_messageId :: Lens' FlowExecutionMessage (Maybe Text)

The unique identifier of the message.

FlowExecutionSummary

flowExecutionSummary_status :: Lens' FlowExecutionSummary (Maybe FlowExecutionStatus)

The current status of the flow execution.

flowExecutionSummary_flowTemplateId :: Lens' FlowExecutionSummary (Maybe Text)

The ID of the flow.

flowExecutionSummary_createdAt :: Lens' FlowExecutionSummary (Maybe UTCTime)

The date and time when the flow execution summary was created.

flowExecutionSummary_flowExecutionId :: Lens' FlowExecutionSummary (Maybe Text)

The ID of the flow execution.

flowExecutionSummary_systemInstanceId :: Lens' FlowExecutionSummary (Maybe Text)

The ID of the system instance that contains the flow.

flowExecutionSummary_updatedAt :: Lens' FlowExecutionSummary (Maybe UTCTime)

The date and time when the flow execution summary was last updated.

FlowTemplateDescription

flowTemplateDescription_summary :: Lens' FlowTemplateDescription (Maybe FlowTemplateSummary)

An object that contains summary information about a workflow.

flowTemplateDescription_definition :: Lens' FlowTemplateDescription (Maybe DefinitionDocument)

A workflow's definition document.

flowTemplateDescription_validatedNamespaceVersion :: Lens' FlowTemplateDescription (Maybe Integer)

The version of the user's namespace against which the workflow was validated. Use this value in your system instance.

FlowTemplateFilter

flowTemplateFilter_name :: Lens' FlowTemplateFilter FlowTemplateFilterName

The name of the search filter field.

flowTemplateFilter_value :: Lens' FlowTemplateFilter [Text]

An array of string values for the search filter field. Multiple values function as AND criteria in the search.

FlowTemplateSummary

flowTemplateSummary_arn :: Lens' FlowTemplateSummary (Maybe Text)

The ARN of the workflow.

flowTemplateSummary_createdAt :: Lens' FlowTemplateSummary (Maybe UTCTime)

The date when the workflow was created.

flowTemplateSummary_revisionNumber :: Lens' FlowTemplateSummary (Maybe Integer)

The revision number of the workflow.

flowTemplateSummary_id :: Lens' FlowTemplateSummary (Maybe Text)

The ID of the workflow.

MetricsConfiguration

metricsConfiguration_cloudMetricEnabled :: Lens' MetricsConfiguration (Maybe Bool)

A Boolean that specifies whether cloud metrics are collected.

metricsConfiguration_metricRuleRoleArn :: Lens' MetricsConfiguration (Maybe Text)

The ARN of the role that is used to collect cloud metrics.

SystemInstanceDescription

systemInstanceDescription_summary :: Lens' SystemInstanceDescription (Maybe SystemInstanceSummary)

An object that contains summary information about a system instance.

systemInstanceDescription_metricsConfiguration :: Lens' SystemInstanceDescription (Maybe MetricsConfiguration)

Undocumented member.

systemInstanceDescription_validatedDependencyRevisions :: Lens' SystemInstanceDescription (Maybe [DependencyRevision])

A list of objects that contain all of the IDs and revision numbers of workflows and systems that are used in a system instance.

systemInstanceDescription_definition :: Lens' SystemInstanceDescription (Maybe DefinitionDocument)

Undocumented member.

systemInstanceDescription_validatedNamespaceVersion :: Lens' SystemInstanceDescription (Maybe Integer)

The version of the user's namespace against which the system instance was validated.

systemInstanceDescription_flowActionsRoleArn :: Lens' SystemInstanceDescription (Maybe Text)

The AWS Identity and Access Management (IAM) role that AWS IoT Things Graph assumes during flow execution in a cloud deployment. This role must have read and write permissionss to AWS Lambda and AWS IoT and to any other AWS services that the flow uses.

systemInstanceDescription_s3BucketName :: Lens' SystemInstanceDescription (Maybe Text)

The Amazon Simple Storage Service bucket where information about a system instance is stored.

SystemInstanceFilter

systemInstanceFilter_value :: Lens' SystemInstanceFilter (Maybe [Text])

An array of string values for the search filter field. Multiple values function as AND criteria in the search.

systemInstanceFilter_name :: Lens' SystemInstanceFilter (Maybe SystemInstanceFilterName)

The name of the search filter field.

SystemInstanceSummary

systemInstanceSummary_status :: Lens' SystemInstanceSummary (Maybe SystemInstanceDeploymentStatus)

The status of the system instance.

systemInstanceSummary_greengrassGroupName :: Lens' SystemInstanceSummary (Maybe Text)

The ID of the Greengrass group where the system instance is deployed.

systemInstanceSummary_arn :: Lens' SystemInstanceSummary (Maybe Text)

The ARN of the system instance.

systemInstanceSummary_createdAt :: Lens' SystemInstanceSummary (Maybe UTCTime)

The date when the system instance was created.

systemInstanceSummary_greengrassGroupId :: Lens' SystemInstanceSummary (Maybe Text)

The ID of the Greengrass group where the system instance is deployed.

systemInstanceSummary_greengrassGroupVersionId :: Lens' SystemInstanceSummary (Maybe Text)

The version of the Greengrass group where the system instance is deployed.

systemInstanceSummary_id :: Lens' SystemInstanceSummary (Maybe Text)

The ID of the system instance.

systemInstanceSummary_updatedAt :: Lens' SystemInstanceSummary (Maybe UTCTime)

The date and time when the system instance was last updated.

systemInstanceSummary_target :: Lens' SystemInstanceSummary (Maybe DeploymentTarget)

The target of the system instance.

SystemTemplateDescription

systemTemplateDescription_summary :: Lens' SystemTemplateDescription (Maybe SystemTemplateSummary)

An object that contains summary information about a system.

systemTemplateDescription_definition :: Lens' SystemTemplateDescription (Maybe DefinitionDocument)

The definition document of a system.

systemTemplateDescription_validatedNamespaceVersion :: Lens' SystemTemplateDescription (Maybe Integer)

The namespace version against which the system was validated. Use this value in your system instance.

SystemTemplateFilter

systemTemplateFilter_name :: Lens' SystemTemplateFilter SystemTemplateFilterName

The name of the system search filter field.

systemTemplateFilter_value :: Lens' SystemTemplateFilter [Text]

An array of string values for the search filter field. Multiple values function as AND criteria in the search.

SystemTemplateSummary

systemTemplateSummary_arn :: Lens' SystemTemplateSummary (Maybe Text)

The ARN of the system.

systemTemplateSummary_createdAt :: Lens' SystemTemplateSummary (Maybe UTCTime)

The date when the system was created.

systemTemplateSummary_revisionNumber :: Lens' SystemTemplateSummary (Maybe Integer)

The revision number of the system.

systemTemplateSummary_id :: Lens' SystemTemplateSummary (Maybe Text)

The ID of the system.

Tag

tag_key :: Lens' Tag Text

The required name of the tag. The string value can be from 1 to 128 Unicode characters in length.

tag_value :: Lens' Tag Text

The optional value of the tag. The string value can be from 1 to 256 Unicode characters in length.

Thing

thing_thingArn :: Lens' Thing (Maybe Text)

The ARN of the thing.

thing_thingName :: Lens' Thing (Maybe Text)

The name of the thing.