Network.AWS.IoTThingsGraph.SearchEntities

Description

Searches for entities of the specified type. You can search for entities in your namespace and the public namespace that you're tracking.

This operation returns paginated results.

Creating a Request

data SearchEntities

See: newSearchEntities smart constructor.

Constructors

SearchEntities'

Fields

filters :: Maybe [EntityFilter]

Optional filter to apply to the search. Valid filters are NAME NAMESPACE, SEMANTIC_TYPE_PATH and REFERENCED_ENTITY_ID. REFERENCED_ENTITY_ID filters on entities that are used by the entity in the result set. For example, you can filter on the ID of a property that is used in a state.

Multiple filters function as OR criteria in the query. Multiple values passed inside the filter function as AND criteria.

namespaceVersion :: Maybe Integer

The version of the user's namespace. Defaults to the latest version of the user's namespace.

nextToken :: Maybe Text

The string that specifies the next page of results. Use this when you're paginating results.

maxResults :: Maybe Natural

The maximum number of results to return in the response.

entityTypes :: [EntityType]

The entity types for which to search.

newSearchEntities :: SearchEntities

Create a value of SearchEntities with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:SearchEntities', searchEntities_filters - Optional filter to apply to the search. Valid filters are NAME NAMESPACE, SEMANTIC_TYPE_PATH and REFERENCED_ENTITY_ID. REFERENCED_ENTITY_ID filters on entities that are used by the entity in the result set. For example, you can filter on the ID of a property that is used in a state.

Multiple filters function as OR criteria in the query. Multiple values passed inside the filter function as AND criteria.

$sel:namespaceVersion:SearchEntities', searchEntities_namespaceVersion - The version of the user's namespace. Defaults to the latest version of the user's namespace.

$sel:nextToken:SearchEntities', searchEntities_nextToken - The string that specifies the next page of results. Use this when you're paginating results.

$sel:maxResults:SearchEntities', searchEntities_maxResults - The maximum number of results to return in the response.

$sel:entityTypes:SearchEntities', searchEntities_entityTypes - The entity types for which to search.

Request Lenses

searchEntities_filters :: Lens' SearchEntities (Maybe [EntityFilter])

Optional filter to apply to the search. Valid filters are NAME NAMESPACE, SEMANTIC_TYPE_PATH and REFERENCED_ENTITY_ID. REFERENCED_ENTITY_ID filters on entities that are used by the entity in the result set. For example, you can filter on the ID of a property that is used in a state.

Multiple filters function as OR criteria in the query. Multiple values passed inside the filter function as AND criteria.

searchEntities_namespaceVersion :: Lens' SearchEntities (Maybe Integer)

The version of the user's namespace. Defaults to the latest version of the user's namespace.

searchEntities_nextToken :: Lens' SearchEntities (Maybe Text)

The string that specifies the next page of results. Use this when you're paginating results.

searchEntities_maxResults :: Lens' SearchEntities (Maybe Natural)

The maximum number of results to return in the response.

searchEntities_entityTypes :: Lens' SearchEntities [EntityType]

The entity types for which to search.

Destructuring the Response

data SearchEntitiesResponse

See: newSearchEntitiesResponse smart constructor.

Constructors

SearchEntitiesResponse'

Fields

nextToken :: Maybe Text

The string to specify as nextToken when you request the next page of results.

descriptions :: Maybe [EntityDescription]

An array of descriptions for each entity returned in the search result.

httpStatus :: Int

The response's http status code.

newSearchEntitiesResponse

Arguments

:: Int
$sel:httpStatus:SearchEntitiesResponse'

-> SearchEntitiesResponse

Create a value of SearchEntitiesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:SearchEntities', searchEntitiesResponse_nextToken - The string to specify as nextToken when you request the next page of results.

$sel:descriptions:SearchEntitiesResponse', searchEntitiesResponse_descriptions - An array of descriptions for each entity returned in the search result.

$sel:httpStatus:SearchEntitiesResponse', searchEntitiesResponse_httpStatus - The response's http status code.

Response Lenses

searchEntitiesResponse_nextToken :: Lens' SearchEntitiesResponse (Maybe Text)

The string to specify as nextToken when you request the next page of results.

searchEntitiesResponse_descriptions :: Lens' SearchEntitiesResponse (Maybe [EntityDescription])

An array of descriptions for each entity returned in the search result.

searchEntitiesResponse_httpStatus :: Lens' SearchEntitiesResponse Int

The response's http status code.