Network.AWS.IoTThingsGraph.Types.FlowTemplateDescription

Description

Documentation

data FlowTemplateDescription

An object that contains a workflow's definition and summary information.

See: newFlowTemplateDescription smart constructor.

Constructors

FlowTemplateDescription'

Fields

summary :: Maybe FlowTemplateSummary

An object that contains summary information about a workflow.

definition :: Maybe DefinitionDocument

A workflow's definition document.

validatedNamespaceVersion :: Maybe Integer

The version of the user's namespace against which the workflow was validated. Use this value in your system instance.

newFlowTemplateDescription :: FlowTemplateDescription

Create a value of FlowTemplateDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:summary:FlowTemplateDescription', flowTemplateDescription_summary - An object that contains summary information about a workflow.

$sel:definition:FlowTemplateDescription', flowTemplateDescription_definition - A workflow's definition document.

$sel:validatedNamespaceVersion:FlowTemplateDescription', flowTemplateDescription_validatedNamespaceVersion - The version of the user's namespace against which the workflow was validated. Use this value in your system instance.

flowTemplateDescription_summary :: Lens' FlowTemplateDescription (Maybe FlowTemplateSummary)

An object that contains summary information about a workflow.

flowTemplateDescription_definition :: Lens' FlowTemplateDescription (Maybe DefinitionDocument)

A workflow's definition document.

flowTemplateDescription_validatedNamespaceVersion :: Lens' FlowTemplateDescription (Maybe Integer)

The version of the user's namespace against which the workflow was validated. Use this value in your system instance.