Network.AWS.IoTThingsGraph.Types.SystemInstanceDescription

Description

Documentation

data SystemInstanceDescription

An object that contains a system instance definition and summary information.

See: newSystemInstanceDescription smart constructor.

Constructors

SystemInstanceDescription'

Fields

summary :: Maybe SystemInstanceSummary

An object that contains summary information about a system instance.

metricsConfiguration :: Maybe MetricsConfiguration

validatedDependencyRevisions :: Maybe [DependencyRevision]

A list of objects that contain all of the IDs and revision numbers of workflows and systems that are used in a system instance.

definition :: Maybe DefinitionDocument

validatedNamespaceVersion :: Maybe Integer

The version of the user's namespace against which the system instance was validated.

flowActionsRoleArn :: Maybe Text

The AWS Identity and Access Management (IAM) role that AWS IoT Things Graph assumes during flow execution in a cloud deployment. This role must have read and write permissionss to AWS Lambda and AWS IoT and to any other AWS services that the flow uses.

s3BucketName :: Maybe Text

The Amazon Simple Storage Service bucket where information about a system instance is stored.

newSystemInstanceDescription :: SystemInstanceDescription

Create a value of SystemInstanceDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:summary:SystemInstanceDescription', systemInstanceDescription_summary - An object that contains summary information about a system instance.

$sel:metricsConfiguration:SystemInstanceDescription', systemInstanceDescription_metricsConfiguration - Undocumented member.

$sel:validatedDependencyRevisions:SystemInstanceDescription', systemInstanceDescription_validatedDependencyRevisions - A list of objects that contain all of the IDs and revision numbers of workflows and systems that are used in a system instance.

$sel:definition:SystemInstanceDescription', systemInstanceDescription_definition - Undocumented member.

$sel:validatedNamespaceVersion:SystemInstanceDescription', systemInstanceDescription_validatedNamespaceVersion - The version of the user's namespace against which the system instance was validated.

$sel:flowActionsRoleArn:SystemInstanceDescription', systemInstanceDescription_flowActionsRoleArn - The AWS Identity and Access Management (IAM) role that AWS IoT Things Graph assumes during flow execution in a cloud deployment. This role must have read and write permissionss to AWS Lambda and AWS IoT and to any other AWS services that the flow uses.

$sel:s3BucketName:SystemInstanceDescription', systemInstanceDescription_s3BucketName - The Amazon Simple Storage Service bucket where information about a system instance is stored.

systemInstanceDescription_summary :: Lens' SystemInstanceDescription (Maybe SystemInstanceSummary)

An object that contains summary information about a system instance.

systemInstanceDescription_metricsConfiguration :: Lens' SystemInstanceDescription (Maybe MetricsConfiguration)

Undocumented member.

systemInstanceDescription_validatedDependencyRevisions :: Lens' SystemInstanceDescription (Maybe [DependencyRevision])

A list of objects that contain all of the IDs and revision numbers of workflows and systems that are used in a system instance.

systemInstanceDescription_definition :: Lens' SystemInstanceDescription (Maybe DefinitionDocument)

Undocumented member.

systemInstanceDescription_validatedNamespaceVersion :: Lens' SystemInstanceDescription (Maybe Integer)

The version of the user's namespace against which the system instance was validated.

systemInstanceDescription_flowActionsRoleArn :: Lens' SystemInstanceDescription (Maybe Text)

The AWS Identity and Access Management (IAM) role that AWS IoT Things Graph assumes during flow execution in a cloud deployment. This role must have read and write permissionss to AWS Lambda and AWS IoT and to any other AWS services that the flow uses.

systemInstanceDescription_s3BucketName :: Lens' SystemInstanceDescription (Maybe Text)

The Amazon Simple Storage Service bucket where information about a system instance is stored.