Network.AWS.IoTThingsGraph.Types.SystemInstanceSummary

Description

Documentation

data SystemInstanceSummary

An object that contains summary information about a system instance.

See: newSystemInstanceSummary smart constructor.

Constructors

SystemInstanceSummary'

Fields

status :: Maybe SystemInstanceDeploymentStatus

The status of the system instance.

greengrassGroupName :: Maybe Text

The ID of the Greengrass group where the system instance is deployed.

arn :: Maybe Text

The ARN of the system instance.

createdAt :: Maybe POSIX

The date when the system instance was created.

greengrassGroupId :: Maybe Text

The ID of the Greengrass group where the system instance is deployed.

greengrassGroupVersionId :: Maybe Text

The version of the Greengrass group where the system instance is deployed.

id :: Maybe Text

The ID of the system instance.

updatedAt :: Maybe POSIX

The date and time when the system instance was last updated.

target :: Maybe DeploymentTarget

The target of the system instance.

newSystemInstanceSummary :: SystemInstanceSummary

Create a value of SystemInstanceSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SystemInstanceSummary', systemInstanceSummary_status - The status of the system instance.

$sel:greengrassGroupName:SystemInstanceSummary', systemInstanceSummary_greengrassGroupName - The ID of the Greengrass group where the system instance is deployed.

$sel:arn:SystemInstanceSummary', systemInstanceSummary_arn - The ARN of the system instance.

$sel:createdAt:SystemInstanceSummary', systemInstanceSummary_createdAt - The date when the system instance was created.

$sel:greengrassGroupId:SystemInstanceSummary', systemInstanceSummary_greengrassGroupId - The ID of the Greengrass group where the system instance is deployed.

$sel:greengrassGroupVersionId:SystemInstanceSummary', systemInstanceSummary_greengrassGroupVersionId - The version of the Greengrass group where the system instance is deployed.

$sel:id:SystemInstanceSummary', systemInstanceSummary_id - The ID of the system instance.

$sel:updatedAt:SystemInstanceSummary', systemInstanceSummary_updatedAt - The date and time when the system instance was last updated.

$sel:target:SystemInstanceSummary', systemInstanceSummary_target - The target of the system instance.

systemInstanceSummary_status :: Lens' SystemInstanceSummary (Maybe SystemInstanceDeploymentStatus)

The status of the system instance.

systemInstanceSummary_greengrassGroupName :: Lens' SystemInstanceSummary (Maybe Text)

The ID of the Greengrass group where the system instance is deployed.

systemInstanceSummary_arn :: Lens' SystemInstanceSummary (Maybe Text)

The ARN of the system instance.

systemInstanceSummary_createdAt :: Lens' SystemInstanceSummary (Maybe UTCTime)

The date when the system instance was created.

systemInstanceSummary_greengrassGroupId :: Lens' SystemInstanceSummary (Maybe Text)

The ID of the Greengrass group where the system instance is deployed.

systemInstanceSummary_greengrassGroupVersionId :: Lens' SystemInstanceSummary (Maybe Text)

The version of the Greengrass group where the system instance is deployed.

systemInstanceSummary_id :: Lens' SystemInstanceSummary (Maybe Text)

The ID of the system instance.

systemInstanceSummary_updatedAt :: Lens' SystemInstanceSummary (Maybe UTCTime)

The date and time when the system instance was last updated.

systemInstanceSummary_target :: Lens' SystemInstanceSummary (Maybe DeploymentTarget)

The target of the system instance.