Network.AWS.IoTThingsGraph.UpdateSystemTemplate

Description

Updates the specified system. You don't need to run this action after updating a workflow. Any deployment that uses the system will see the changes in the system when it is redeployed.

Creating a Request

data UpdateSystemTemplate

See: newUpdateSystemTemplate smart constructor.

Constructors

UpdateSystemTemplate'

Fields

compatibleNamespaceVersion :: Maybe Integer

The version of the user's namespace. Defaults to the latest version of the user's namespace.

If no value is specified, the latest version is used by default.

id :: Text

The ID of the system to be updated.

The ID should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:system:SYSTEMNAME

definition :: DefinitionDocument

The DefinitionDocument that contains the updated system definition.

newUpdateSystemTemplate

Arguments

:: Text
$sel:id:UpdateSystemTemplate'

-> DefinitionDocument
$sel:definition:UpdateSystemTemplate'

-> UpdateSystemTemplate

Create a value of UpdateSystemTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:compatibleNamespaceVersion:UpdateSystemTemplate', updateSystemTemplate_compatibleNamespaceVersion - The version of the user's namespace. Defaults to the latest version of the user's namespace.

If no value is specified, the latest version is used by default.

$sel:id:UpdateSystemTemplate', updateSystemTemplate_id - The ID of the system to be updated.

The ID should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:system:SYSTEMNAME

$sel:definition:UpdateSystemTemplate', updateSystemTemplate_definition - The DefinitionDocument that contains the updated system definition.

Request Lenses

updateSystemTemplate_compatibleNamespaceVersion :: Lens' UpdateSystemTemplate (Maybe Integer)

The version of the user's namespace. Defaults to the latest version of the user's namespace.

If no value is specified, the latest version is used by default.

updateSystemTemplate_id :: Lens' UpdateSystemTemplate Text

The ID of the system to be updated.

The ID should be in the following format.

urn:tdm:REGION/ACCOUNT ID/default:system:SYSTEMNAME

updateSystemTemplate_definition :: Lens' UpdateSystemTemplate DefinitionDocument

The DefinitionDocument that contains the updated system definition.

Destructuring the Response

data UpdateSystemTemplateResponse

See: newUpdateSystemTemplateResponse smart constructor.

Constructors

UpdateSystemTemplateResponse'

Fields

summary :: Maybe SystemTemplateSummary

An object containing summary information about the updated system.

httpStatus :: Int

The response's http status code.

newUpdateSystemTemplateResponse

Arguments

:: Int
$sel:httpStatus:UpdateSystemTemplateResponse'

-> UpdateSystemTemplateResponse

Create a value of UpdateSystemTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:summary:UpdateSystemTemplateResponse', updateSystemTemplateResponse_summary - An object containing summary information about the updated system.

$sel:httpStatus:UpdateSystemTemplateResponse', updateSystemTemplateResponse_httpStatus - The response's http status code.

Response Lenses

updateSystemTemplateResponse_summary :: Lens' UpdateSystemTemplateResponse (Maybe SystemTemplateSummary)

An object containing summary information about the updated system.

updateSystemTemplateResponse_httpStatus :: Lens' UpdateSystemTemplateResponse Int

The response's http status code.