Network.AWS.IoTThingsGraph.UploadEntityDefinitions

Description

Asynchronously uploads one or more entity definitions to the user's namespace. The document parameter is required if syncWithPublicNamespace and deleteExistingEntites are false. If the syncWithPublicNamespace parameter is set to true, the user's namespace will synchronize with the latest version of the public namespace. If deprecateExistingEntities is set to true, all entities in the latest version will be deleted before the new DefinitionDocument is uploaded.

When a user uploads entity definitions for the first time, the service creates a new namespace for the user. The new namespace tracks the public namespace. Currently users can have only one namespace. The namespace version increments whenever a user uploads entity definitions that are backwards-incompatible and whenever a user sets the syncWithPublicNamespace parameter or the deprecateExistingEntities parameter to true.

The IDs for all of the entities should be in URN format. Each entity must be in the user's namespace. Users can't create entities in the public namespace, but entity definitions can refer to entities in the public namespace.

Valid entities are Device, DeviceModel, Service, Capability, State, Action, Event, Property, Mapping, Enum.

Creating a Request

data UploadEntityDefinitions

See: newUploadEntityDefinitions smart constructor.

Constructors

UploadEntityDefinitions'

Fields

syncWithPublicNamespace :: Maybe Bool

A Boolean that specifies whether to synchronize with the latest version of the public namespace. If set to true, the upload will create a new namespace version.

deprecateExistingEntities :: Maybe Bool

A Boolean that specifies whether to deprecate all entities in the latest version before uploading the new DefinitionDocument. If set to true, the upload will create a new namespace version.

document :: Maybe DefinitionDocument

The DefinitionDocument that defines the updated entities.

newUploadEntityDefinitions :: UploadEntityDefinitions

Create a value of UploadEntityDefinitions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:syncWithPublicNamespace:UploadEntityDefinitions', uploadEntityDefinitions_syncWithPublicNamespace - A Boolean that specifies whether to synchronize with the latest version of the public namespace. If set to true, the upload will create a new namespace version.

$sel:deprecateExistingEntities:UploadEntityDefinitions', uploadEntityDefinitions_deprecateExistingEntities - A Boolean that specifies whether to deprecate all entities in the latest version before uploading the new DefinitionDocument. If set to true, the upload will create a new namespace version.

$sel:document:UploadEntityDefinitions', uploadEntityDefinitions_document - The DefinitionDocument that defines the updated entities.

Request Lenses

uploadEntityDefinitions_syncWithPublicNamespace :: Lens' UploadEntityDefinitions (Maybe Bool)

A Boolean that specifies whether to synchronize with the latest version of the public namespace. If set to true, the upload will create a new namespace version.

uploadEntityDefinitions_deprecateExistingEntities :: Lens' UploadEntityDefinitions (Maybe Bool)

A Boolean that specifies whether to deprecate all entities in the latest version before uploading the new DefinitionDocument. If set to true, the upload will create a new namespace version.

uploadEntityDefinitions_document :: Lens' UploadEntityDefinitions (Maybe DefinitionDocument)

The DefinitionDocument that defines the updated entities.

Destructuring the Response

data UploadEntityDefinitionsResponse

See: newUploadEntityDefinitionsResponse smart constructor.

Constructors

UploadEntityDefinitionsResponse'

Fields

httpStatus :: Int

The response's http status code.

uploadId :: Text

The ID that specifies the upload action. You can use this to track the status of the upload.

newUploadEntityDefinitionsResponse

Arguments

:: Int
$sel:httpStatus:UploadEntityDefinitionsResponse'

-> Text
$sel:uploadId:UploadEntityDefinitionsResponse'

-> UploadEntityDefinitionsResponse

Create a value of UploadEntityDefinitionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UploadEntityDefinitionsResponse', uploadEntityDefinitionsResponse_httpStatus - The response's http status code.

$sel:uploadId:UploadEntityDefinitionsResponse', uploadEntityDefinitionsResponse_uploadId - The ID that specifies the upload action. You can use this to track the status of the upload.

Response Lenses

uploadEntityDefinitionsResponse_httpStatus :: Lens' UploadEntityDefinitionsResponse Int

The response's http status code.

uploadEntityDefinitionsResponse_uploadId :: Lens' UploadEntityDefinitionsResponse Text

The ID that specifies the upload action. You can use this to track the status of the upload.