Network.AWS.IoTWireless.AssociateAwsAccountWithPartnerAccount

Description

Associates a partner account with your AWS account.

Creating a Request

data AssociateAwsAccountWithPartnerAccount

See: newAssociateAwsAccountWithPartnerAccount smart constructor.

Constructors

AssociateAwsAccountWithPartnerAccount'

Fields

clientRequestToken :: Maybe Text

Each resource must have a unique client request token. If you try to create a new resource with the same token as a resource that already exists, an exception occurs. If you omit this value, AWS SDKs will automatically generate a unique client request.

tags :: Maybe [Tag]

The tags to attach to the specified resource. Tags are metadata that you can use to manage a resource.

sidewalk :: SidewalkAccountInfo

The Sidewalk account credentials.

newAssociateAwsAccountWithPartnerAccount

Arguments

:: SidewalkAccountInfo
$sel:sidewalk:AssociateAwsAccountWithPartnerAccount'

-> AssociateAwsAccountWithPartnerAccount

Create a value of AssociateAwsAccountWithPartnerAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientRequestToken:AssociateAwsAccountWithPartnerAccount', associateAwsAccountWithPartnerAccount_clientRequestToken - Each resource must have a unique client request token. If you try to create a new resource with the same token as a resource that already exists, an exception occurs. If you omit this value, AWS SDKs will automatically generate a unique client request.

$sel:tags:AssociateAwsAccountWithPartnerAccount', associateAwsAccountWithPartnerAccount_tags - The tags to attach to the specified resource. Tags are metadata that you can use to manage a resource.

$sel:sidewalk:AssociateAwsAccountWithPartnerAccount', associateAwsAccountWithPartnerAccount_sidewalk - The Sidewalk account credentials.

Request Lenses

associateAwsAccountWithPartnerAccount_clientRequestToken :: Lens' AssociateAwsAccountWithPartnerAccount (Maybe Text)

Each resource must have a unique client request token. If you try to create a new resource with the same token as a resource that already exists, an exception occurs. If you omit this value, AWS SDKs will automatically generate a unique client request.

associateAwsAccountWithPartnerAccount_tags :: Lens' AssociateAwsAccountWithPartnerAccount (Maybe [Tag])

The tags to attach to the specified resource. Tags are metadata that you can use to manage a resource.

associateAwsAccountWithPartnerAccount_sidewalk :: Lens' AssociateAwsAccountWithPartnerAccount SidewalkAccountInfo

The Sidewalk account credentials.

Destructuring the Response

data AssociateAwsAccountWithPartnerAccountResponse

See: newAssociateAwsAccountWithPartnerAccountResponse smart constructor.

Constructors

AssociateAwsAccountWithPartnerAccountResponse'

Fields

sidewalk :: Maybe SidewalkAccountInfo

The Sidewalk account credentials.

arn :: Maybe Text

The Amazon Resource Name of the resource.

httpStatus :: Int

The response's http status code.

newAssociateAwsAccountWithPartnerAccountResponse

Arguments

:: Int
$sel:httpStatus:AssociateAwsAccountWithPartnerAccountResponse'

-> AssociateAwsAccountWithPartnerAccountResponse

Create a value of AssociateAwsAccountWithPartnerAccountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sidewalk:AssociateAwsAccountWithPartnerAccount', associateAwsAccountWithPartnerAccountResponse_sidewalk - The Sidewalk account credentials.

$sel:arn:AssociateAwsAccountWithPartnerAccountResponse', associateAwsAccountWithPartnerAccountResponse_arn - The Amazon Resource Name of the resource.

$sel:httpStatus:AssociateAwsAccountWithPartnerAccountResponse', associateAwsAccountWithPartnerAccountResponse_httpStatus - The response's http status code.

Response Lenses

associateAwsAccountWithPartnerAccountResponse_sidewalk :: Lens' AssociateAwsAccountWithPartnerAccountResponse (Maybe SidewalkAccountInfo)

The Sidewalk account credentials.

associateAwsAccountWithPartnerAccountResponse_arn :: Lens' AssociateAwsAccountWithPartnerAccountResponse (Maybe Text)

The Amazon Resource Name of the resource.

associateAwsAccountWithPartnerAccountResponse_httpStatus :: Lens' AssociateAwsAccountWithPartnerAccountResponse Int

The response's http status code.