Network.AWS.IoTWireless.CreateWirelessDevice

Description

Provisions a wireless device.

Creating a Request

data CreateWirelessDevice

See: newCreateWirelessDevice smart constructor.

Constructors

CreateWirelessDevice'

Fields

loRaWAN :: Maybe LoRaWANDevice

The device configuration information to use to create the wireless device.

name :: Maybe Text

The name of the new resource.

clientRequestToken :: Maybe Text

Each resource must have a unique client request token. If you try to create a new resource with the same token as a resource that already exists, an exception occurs. If you omit this value, AWS SDKs will automatically generate a unique client request.

description :: Maybe Text

The description of the new resource.

tags :: Maybe [Tag]

The tags to attach to the new wireless device. Tags are metadata that you can use to manage a resource.

type' :: WirelessDeviceType

The wireless device type.

destinationName :: Text

The name of the destination to assign to the new wireless device.

newCreateWirelessDevice

Arguments

:: WirelessDeviceType
$sel:type':CreateWirelessDevice'

-> Text
$sel:destinationName:CreateWirelessDevice'

-> CreateWirelessDevice

Create a value of CreateWirelessDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loRaWAN:CreateWirelessDevice', createWirelessDevice_loRaWAN - The device configuration information to use to create the wireless device.

$sel:name:CreateWirelessDevice', createWirelessDevice_name - The name of the new resource.

$sel:clientRequestToken:CreateWirelessDevice', createWirelessDevice_clientRequestToken - Each resource must have a unique client request token. If you try to create a new resource with the same token as a resource that already exists, an exception occurs. If you omit this value, AWS SDKs will automatically generate a unique client request.

$sel:description:CreateWirelessDevice', createWirelessDevice_description - The description of the new resource.

$sel:tags:CreateWirelessDevice', createWirelessDevice_tags - The tags to attach to the new wireless device. Tags are metadata that you can use to manage a resource.

$sel:type':CreateWirelessDevice', createWirelessDevice_type - The wireless device type.

$sel:destinationName:CreateWirelessDevice', createWirelessDevice_destinationName - The name of the destination to assign to the new wireless device.

Request Lenses

createWirelessDevice_loRaWAN :: Lens' CreateWirelessDevice (Maybe LoRaWANDevice)

The device configuration information to use to create the wireless device.

createWirelessDevice_name :: Lens' CreateWirelessDevice (Maybe Text)

The name of the new resource.

createWirelessDevice_clientRequestToken :: Lens' CreateWirelessDevice (Maybe Text)

Each resource must have a unique client request token. If you try to create a new resource with the same token as a resource that already exists, an exception occurs. If you omit this value, AWS SDKs will automatically generate a unique client request.

createWirelessDevice_description :: Lens' CreateWirelessDevice (Maybe Text)

The description of the new resource.

createWirelessDevice_tags :: Lens' CreateWirelessDevice (Maybe [Tag])

The tags to attach to the new wireless device. Tags are metadata that you can use to manage a resource.

createWirelessDevice_type :: Lens' CreateWirelessDevice WirelessDeviceType

The wireless device type.

createWirelessDevice_destinationName :: Lens' CreateWirelessDevice Text

The name of the destination to assign to the new wireless device.

Destructuring the Response

data CreateWirelessDeviceResponse

See: newCreateWirelessDeviceResponse smart constructor.

Constructors

CreateWirelessDeviceResponse'

Fields

arn :: Maybe Text

The Amazon Resource Name of the new resource.

id :: Maybe Text

The ID of the new wireless device.

httpStatus :: Int

The response's http status code.

newCreateWirelessDeviceResponse

Arguments

:: Int
$sel:httpStatus:CreateWirelessDeviceResponse'

-> CreateWirelessDeviceResponse

Create a value of CreateWirelessDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:CreateWirelessDeviceResponse', createWirelessDeviceResponse_arn - The Amazon Resource Name of the new resource.

$sel:id:CreateWirelessDeviceResponse', createWirelessDeviceResponse_id - The ID of the new wireless device.

$sel:httpStatus:CreateWirelessDeviceResponse', createWirelessDeviceResponse_httpStatus - The response's http status code.

Response Lenses

createWirelessDeviceResponse_arn :: Lens' CreateWirelessDeviceResponse (Maybe Text)

The Amazon Resource Name of the new resource.

createWirelessDeviceResponse_id :: Lens' CreateWirelessDeviceResponse (Maybe Text)

The ID of the new wireless device.

createWirelessDeviceResponse_httpStatus :: Lens' CreateWirelessDeviceResponse Int

The response's http status code.