Network.AWS.IoTWireless.GetDestination

Description

Gets information about a destination.

Creating a Request

data GetDestination

See: newGetDestination smart constructor.

Constructors

GetDestination'

Fields

name :: Text

The name of the resource to get.

newGetDestination

Arguments

:: Text
$sel:name:GetDestination'

-> GetDestination

Create a value of GetDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetDestination', getDestination_name - The name of the resource to get.

Request Lenses

getDestination_name :: Lens' GetDestination Text

The name of the resource to get.

Destructuring the Response

data GetDestinationResponse

See: newGetDestinationResponse smart constructor.

Constructors

GetDestinationResponse'

Fields

expressionType :: Maybe ExpressionType

The type of value in Expression.

arn :: Maybe Text

The Amazon Resource Name of the resource.

name :: Maybe Text

The name of the resource.

expression :: Maybe Text

The rule name or topic rule to send messages to.

description :: Maybe Text

The description of the resource.

roleArn :: Maybe Text

The ARN of the IAM Role that authorizes the destination.

httpStatus :: Int

The response's http status code.

newGetDestinationResponse

Arguments

:: Int
$sel:httpStatus:GetDestinationResponse'

-> GetDestinationResponse

Create a value of GetDestinationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expressionType:GetDestinationResponse', getDestinationResponse_expressionType - The type of value in Expression.

$sel:arn:GetDestinationResponse', getDestinationResponse_arn - The Amazon Resource Name of the resource.

$sel:name:GetDestination', getDestinationResponse_name - The name of the resource.

$sel:expression:GetDestinationResponse', getDestinationResponse_expression - The rule name or topic rule to send messages to.

$sel:description:GetDestinationResponse', getDestinationResponse_description - The description of the resource.

$sel:roleArn:GetDestinationResponse', getDestinationResponse_roleArn - The ARN of the IAM Role that authorizes the destination.

$sel:httpStatus:GetDestinationResponse', getDestinationResponse_httpStatus - The response's http status code.

Response Lenses

getDestinationResponse_expressionType :: Lens' GetDestinationResponse (Maybe ExpressionType)

The type of value in Expression.

getDestinationResponse_arn :: Lens' GetDestinationResponse (Maybe Text)

The Amazon Resource Name of the resource.

getDestinationResponse_name :: Lens' GetDestinationResponse (Maybe Text)

The name of the resource.

getDestinationResponse_expression :: Lens' GetDestinationResponse (Maybe Text)

The rule name or topic rule to send messages to.

getDestinationResponse_description :: Lens' GetDestinationResponse (Maybe Text)

The description of the resource.

getDestinationResponse_roleArn :: Lens' GetDestinationResponse (Maybe Text)

The ARN of the IAM Role that authorizes the destination.

getDestinationResponse_httpStatus :: Lens' GetDestinationResponse Int

The response's http status code.