Network.AWS.IoTWireless.GetResourceLogLevel

Description

Fetches the log-level override, if any, for a given resource-ID and resource-type. It can be used for a wireless device or a wireless gateway.

Creating a Request

data GetResourceLogLevel

See: newGetResourceLogLevel smart constructor.

Constructors

GetResourceLogLevel'

Fields

resourceIdentifier :: Text

resourceType :: Text

The type of the resource, which can be WirelessDevice or WirelessGateway.

newGetResourceLogLevel

Arguments

:: Text
$sel:resourceIdentifier:GetResourceLogLevel'

-> Text
$sel:resourceType:GetResourceLogLevel'

-> GetResourceLogLevel

Create a value of GetResourceLogLevel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceIdentifier:GetResourceLogLevel', getResourceLogLevel_resourceIdentifier - Undocumented member.

$sel:resourceType:GetResourceLogLevel', getResourceLogLevel_resourceType - The type of the resource, which can be WirelessDevice or WirelessGateway.

Request Lenses

getResourceLogLevel_resourceIdentifier :: Lens' GetResourceLogLevel Text

Undocumented member.

getResourceLogLevel_resourceType :: Lens' GetResourceLogLevel Text

The type of the resource, which can be WirelessDevice or WirelessGateway.

Destructuring the Response

data GetResourceLogLevelResponse

See: newGetResourceLogLevelResponse smart constructor.

Constructors

GetResourceLogLevelResponse'

Fields

logLevel :: Maybe LogLevel

httpStatus :: Int

The response's http status code.

newGetResourceLogLevelResponse

Arguments

:: Int
$sel:httpStatus:GetResourceLogLevelResponse'

-> GetResourceLogLevelResponse

Create a value of GetResourceLogLevelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logLevel:GetResourceLogLevelResponse', getResourceLogLevelResponse_logLevel - Undocumented member.

$sel:httpStatus:GetResourceLogLevelResponse', getResourceLogLevelResponse_httpStatus - The response's http status code.

Response Lenses

getResourceLogLevelResponse_logLevel :: Lens' GetResourceLogLevelResponse (Maybe LogLevel)

Undocumented member.

getResourceLogLevelResponse_httpStatus :: Lens' GetResourceLogLevelResponse Int

The response's http status code.