Network.AWS.IoTWireless.GetWirelessGatewayStatistics

Description

Gets operating information about a wireless gateway.

Creating a Request

data GetWirelessGatewayStatistics

See: newGetWirelessGatewayStatistics smart constructor.

Constructors

GetWirelessGatewayStatistics'

Fields

wirelessGatewayId :: Text

The ID of the wireless gateway for which to get the data.

newGetWirelessGatewayStatistics

Arguments

:: Text
$sel:wirelessGatewayId:GetWirelessGatewayStatistics'

-> GetWirelessGatewayStatistics

Create a value of GetWirelessGatewayStatistics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:wirelessGatewayId:GetWirelessGatewayStatistics', getWirelessGatewayStatistics_wirelessGatewayId - The ID of the wireless gateway for which to get the data.

Request Lenses

getWirelessGatewayStatistics_wirelessGatewayId :: Lens' GetWirelessGatewayStatistics Text

The ID of the wireless gateway for which to get the data.

Destructuring the Response

data GetWirelessGatewayStatisticsResponse

See: newGetWirelessGatewayStatisticsResponse smart constructor.

Constructors

GetWirelessGatewayStatisticsResponse'

Fields

connectionStatus :: Maybe ConnectionStatus

The connection status of the wireless gateway.

lastUplinkReceivedAt :: Maybe Text

The date and time when the most recent uplink was received.

wirelessGatewayId :: Maybe Text

The ID of the wireless gateway.

httpStatus :: Int

The response's http status code.

newGetWirelessGatewayStatisticsResponse

Arguments

:: Int
$sel:httpStatus:GetWirelessGatewayStatisticsResponse'

-> GetWirelessGatewayStatisticsResponse

Create a value of GetWirelessGatewayStatisticsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectionStatus:GetWirelessGatewayStatisticsResponse', getWirelessGatewayStatisticsResponse_connectionStatus - The connection status of the wireless gateway.

$sel:lastUplinkReceivedAt:GetWirelessGatewayStatisticsResponse', getWirelessGatewayStatisticsResponse_lastUplinkReceivedAt - The date and time when the most recent uplink was received.

$sel:wirelessGatewayId:GetWirelessGatewayStatistics', getWirelessGatewayStatisticsResponse_wirelessGatewayId - The ID of the wireless gateway.

$sel:httpStatus:GetWirelessGatewayStatisticsResponse', getWirelessGatewayStatisticsResponse_httpStatus - The response's http status code.

Response Lenses

getWirelessGatewayStatisticsResponse_connectionStatus :: Lens' GetWirelessGatewayStatisticsResponse (Maybe ConnectionStatus)

The connection status of the wireless gateway.

getWirelessGatewayStatisticsResponse_lastUplinkReceivedAt :: Lens' GetWirelessGatewayStatisticsResponse (Maybe Text)

The date and time when the most recent uplink was received.

getWirelessGatewayStatisticsResponse_wirelessGatewayId :: Lens' GetWirelessGatewayStatisticsResponse (Maybe Text)

The ID of the wireless gateway.

getWirelessGatewayStatisticsResponse_httpStatus :: Lens' GetWirelessGatewayStatisticsResponse Int

The response's http status code.