Network.AWS.IoTWireless.ListWirelessDevices

Description

Lists the wireless devices registered to your AWS account.

Creating a Request

data ListWirelessDevices

See: newListWirelessDevices smart constructor.

Constructors

ListWirelessDevices'

Fields

serviceProfileId :: Maybe Text

A filter to list only the wireless devices that use this service profile.

deviceProfileId :: Maybe Text

A filter to list only the wireless devices that use this device profile.

nextToken :: Maybe Text

To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

wirelessDeviceType :: Maybe WirelessDeviceType

A filter to list only the wireless devices that use this wireless device type.

destinationName :: Maybe Text

A filter to list only the wireless devices that use this destination.

maxResults :: Maybe Natural

The maximum number of results to return in this operation.

newListWirelessDevices :: ListWirelessDevices

Create a value of ListWirelessDevices with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceProfileId:ListWirelessDevices', listWirelessDevices_serviceProfileId - A filter to list only the wireless devices that use this service profile.

$sel:deviceProfileId:ListWirelessDevices', listWirelessDevices_deviceProfileId - A filter to list only the wireless devices that use this device profile.

$sel:nextToken:ListWirelessDevices', listWirelessDevices_nextToken - To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

$sel:wirelessDeviceType:ListWirelessDevices', listWirelessDevices_wirelessDeviceType - A filter to list only the wireless devices that use this wireless device type.

$sel:destinationName:ListWirelessDevices', listWirelessDevices_destinationName - A filter to list only the wireless devices that use this destination.

$sel:maxResults:ListWirelessDevices', listWirelessDevices_maxResults - The maximum number of results to return in this operation.

Request Lenses

listWirelessDevices_serviceProfileId :: Lens' ListWirelessDevices (Maybe Text)

A filter to list only the wireless devices that use this service profile.

listWirelessDevices_deviceProfileId :: Lens' ListWirelessDevices (Maybe Text)

A filter to list only the wireless devices that use this device profile.

listWirelessDevices_nextToken :: Lens' ListWirelessDevices (Maybe Text)

To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

listWirelessDevices_wirelessDeviceType :: Lens' ListWirelessDevices (Maybe WirelessDeviceType)

A filter to list only the wireless devices that use this wireless device type.

listWirelessDevices_destinationName :: Lens' ListWirelessDevices (Maybe Text)

A filter to list only the wireless devices that use this destination.

listWirelessDevices_maxResults :: Lens' ListWirelessDevices (Maybe Natural)

The maximum number of results to return in this operation.

Destructuring the Response

data ListWirelessDevicesResponse

See: newListWirelessDevicesResponse smart constructor.

Constructors

ListWirelessDevicesResponse'

Fields

wirelessDeviceList :: Maybe [WirelessDeviceStatistics]

The ID of the wireless device.

nextToken :: Maybe Text

The token to use to get the next set of results, or null if there are no additional results.

httpStatus :: Int

The response's http status code.

newListWirelessDevicesResponse

Arguments

:: Int
$sel:httpStatus:ListWirelessDevicesResponse'

-> ListWirelessDevicesResponse

Create a value of ListWirelessDevicesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:wirelessDeviceList:ListWirelessDevicesResponse', listWirelessDevicesResponse_wirelessDeviceList - The ID of the wireless device.

$sel:nextToken:ListWirelessDevices', listWirelessDevicesResponse_nextToken - The token to use to get the next set of results, or null if there are no additional results.

$sel:httpStatus:ListWirelessDevicesResponse', listWirelessDevicesResponse_httpStatus - The response's http status code.

Response Lenses

listWirelessDevicesResponse_wirelessDeviceList :: Lens' ListWirelessDevicesResponse (Maybe [WirelessDeviceStatistics])

The ID of the wireless device.

listWirelessDevicesResponse_nextToken :: Lens' ListWirelessDevicesResponse (Maybe Text)

The token to use to get the next set of results, or null if there are no additional results.

listWirelessDevicesResponse_httpStatus :: Lens' ListWirelessDevicesResponse Int

The response's http status code.