Network.AWS.IoTWireless.Types.Destinations

Description

Documentation

data Destinations

Describes a destination.

See: newDestinations smart constructor.

Constructors

Destinations'

Fields

expressionType :: Maybe ExpressionType

The type of value in Expression.

arn :: Maybe Text

The Amazon Resource Name of the resource.

name :: Maybe Text

The name of the resource.

expression :: Maybe Text

The rule name or topic rule to send messages to.

description :: Maybe Text

The description of the resource.

roleArn :: Maybe Text

The ARN of the IAM Role that authorizes the destination.

newDestinations :: Destinations

Create a value of Destinations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expressionType:Destinations', destinations_expressionType - The type of value in Expression.

$sel:arn:Destinations', destinations_arn - The Amazon Resource Name of the resource.

$sel:name:Destinations', destinations_name - The name of the resource.

$sel:expression:Destinations', destinations_expression - The rule name or topic rule to send messages to.

$sel:description:Destinations', destinations_description - The description of the resource.

$sel:roleArn:Destinations', destinations_roleArn - The ARN of the IAM Role that authorizes the destination.

destinations_expressionType :: Lens' Destinations (Maybe ExpressionType)

The type of value in Expression.

destinations_arn :: Lens' Destinations (Maybe Text)

The Amazon Resource Name of the resource.

destinations_name :: Lens' Destinations (Maybe Text)

The name of the resource.

destinations_expression :: Lens' Destinations (Maybe Text)

The rule name or topic rule to send messages to.

destinations_description :: Lens' Destinations (Maybe Text)

The description of the resource.

destinations_roleArn :: Lens' Destinations (Maybe Text)

The ARN of the IAM Role that authorizes the destination.