Network.AWS.IoTWireless.Types.LoRaWANDeviceProfile

Description

Documentation

data LoRaWANDeviceProfile

LoRaWANDeviceProfile object.

See: newLoRaWANDeviceProfile smart constructor.

Constructors

LoRaWANDeviceProfile'

Fields

rfRegion :: Maybe Text

The frequency band (RFRegion) value.

pingSlotFreq :: Maybe Natural

The PingSlotFreq value.

classBTimeout :: Maybe Natural

The ClassBTimeout value.

maxEirp :: Maybe Natural

The MaxEIRP value.

rxDataRate2 :: Maybe Natural

The RXDataRate2 value.

macVersion :: Maybe Text

The MAC version (such as OTAA 1.1 or OTAA 1.0.3) to use with this device profile.

factoryPresetFreqsList :: Maybe [Natural]

The list of values that make up the FactoryPresetFreqs value.

rxFreq2 :: Maybe Natural

The RXFreq2 value.

classCTimeout :: Maybe Natural

The ClassCTimeout value.

regParamsRevision :: Maybe Text

The version of regional parameters.

pingSlotDr :: Maybe Natural

The PingSlotDR value.

supports32BitFCnt :: Maybe Bool

The Supports32BitFCnt value.

maxDutyCycle :: Maybe Natural

The MaxDutyCycle value.

rxDrOffset1 :: Maybe Natural

The RXDROffset1 value.

supportsClassB :: Maybe Bool

The SupportsClassB value.

supportsJoin :: Maybe Bool

The SupportsJoin value.

rxDelay1 :: Maybe Natural

The RXDelay1 value.

supportsClassC :: Maybe Bool

The SupportsClassC value.

pingSlotPeriod :: Maybe Natural

The PingSlotPeriod value.

newLoRaWANDeviceProfile :: LoRaWANDeviceProfile

Create a value of LoRaWANDeviceProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rfRegion:LoRaWANDeviceProfile', loRaWANDeviceProfile_rfRegion - The frequency band (RFRegion) value.

$sel:pingSlotFreq:LoRaWANDeviceProfile', loRaWANDeviceProfile_pingSlotFreq - The PingSlotFreq value.

$sel:classBTimeout:LoRaWANDeviceProfile', loRaWANDeviceProfile_classBTimeout - The ClassBTimeout value.

$sel:maxEirp:LoRaWANDeviceProfile', loRaWANDeviceProfile_maxEirp - The MaxEIRP value.

$sel:rxDataRate2:LoRaWANDeviceProfile', loRaWANDeviceProfile_rxDataRate2 - The RXDataRate2 value.

$sel:macVersion:LoRaWANDeviceProfile', loRaWANDeviceProfile_macVersion - The MAC version (such as OTAA 1.1 or OTAA 1.0.3) to use with this device profile.

$sel:factoryPresetFreqsList:LoRaWANDeviceProfile', loRaWANDeviceProfile_factoryPresetFreqsList - The list of values that make up the FactoryPresetFreqs value.

$sel:rxFreq2:LoRaWANDeviceProfile', loRaWANDeviceProfile_rxFreq2 - The RXFreq2 value.

$sel:classCTimeout:LoRaWANDeviceProfile', loRaWANDeviceProfile_classCTimeout - The ClassCTimeout value.

$sel:regParamsRevision:LoRaWANDeviceProfile', loRaWANDeviceProfile_regParamsRevision - The version of regional parameters.

$sel:pingSlotDr:LoRaWANDeviceProfile', loRaWANDeviceProfile_pingSlotDr - The PingSlotDR value.

$sel:supports32BitFCnt:LoRaWANDeviceProfile', loRaWANDeviceProfile_supports32BitFCnt - The Supports32BitFCnt value.

$sel:maxDutyCycle:LoRaWANDeviceProfile', loRaWANDeviceProfile_maxDutyCycle - The MaxDutyCycle value.

$sel:rxDrOffset1:LoRaWANDeviceProfile', loRaWANDeviceProfile_rxDrOffset1 - The RXDROffset1 value.

$sel:supportsClassB:LoRaWANDeviceProfile', loRaWANDeviceProfile_supportsClassB - The SupportsClassB value.

$sel:supportsJoin:LoRaWANDeviceProfile', loRaWANDeviceProfile_supportsJoin - The SupportsJoin value.

$sel:rxDelay1:LoRaWANDeviceProfile', loRaWANDeviceProfile_rxDelay1 - The RXDelay1 value.

$sel:supportsClassC:LoRaWANDeviceProfile', loRaWANDeviceProfile_supportsClassC - The SupportsClassC value.

$sel:pingSlotPeriod:LoRaWANDeviceProfile', loRaWANDeviceProfile_pingSlotPeriod - The PingSlotPeriod value.

loRaWANDeviceProfile_rfRegion :: Lens' LoRaWANDeviceProfile (Maybe Text)

The frequency band (RFRegion) value.

loRaWANDeviceProfile_pingSlotFreq :: Lens' LoRaWANDeviceProfile (Maybe Natural)

The PingSlotFreq value.

loRaWANDeviceProfile_classBTimeout :: Lens' LoRaWANDeviceProfile (Maybe Natural)

The ClassBTimeout value.

loRaWANDeviceProfile_maxEirp :: Lens' LoRaWANDeviceProfile (Maybe Natural)

The MaxEIRP value.

loRaWANDeviceProfile_rxDataRate2 :: Lens' LoRaWANDeviceProfile (Maybe Natural)

The RXDataRate2 value.

loRaWANDeviceProfile_macVersion :: Lens' LoRaWANDeviceProfile (Maybe Text)

The MAC version (such as OTAA 1.1 or OTAA 1.0.3) to use with this device profile.

loRaWANDeviceProfile_factoryPresetFreqsList :: Lens' LoRaWANDeviceProfile (Maybe [Natural])

The list of values that make up the FactoryPresetFreqs value.

loRaWANDeviceProfile_rxFreq2 :: Lens' LoRaWANDeviceProfile (Maybe Natural)

The RXFreq2 value.

loRaWANDeviceProfile_classCTimeout :: Lens' LoRaWANDeviceProfile (Maybe Natural)

The ClassCTimeout value.

loRaWANDeviceProfile_regParamsRevision :: Lens' LoRaWANDeviceProfile (Maybe Text)

The version of regional parameters.

loRaWANDeviceProfile_pingSlotDr :: Lens' LoRaWANDeviceProfile (Maybe Natural)

The PingSlotDR value.

loRaWANDeviceProfile_supports32BitFCnt :: Lens' LoRaWANDeviceProfile (Maybe Bool)

The Supports32BitFCnt value.

loRaWANDeviceProfile_maxDutyCycle :: Lens' LoRaWANDeviceProfile (Maybe Natural)

The MaxDutyCycle value.

loRaWANDeviceProfile_rxDrOffset1 :: Lens' LoRaWANDeviceProfile (Maybe Natural)

The RXDROffset1 value.

loRaWANDeviceProfile_supportsClassB :: Lens' LoRaWANDeviceProfile (Maybe Bool)

The SupportsClassB value.

loRaWANDeviceProfile_supportsJoin :: Lens' LoRaWANDeviceProfile (Maybe Bool)

The SupportsJoin value.

loRaWANDeviceProfile_rxDelay1 :: Lens' LoRaWANDeviceProfile (Maybe Natural)

The RXDelay1 value.

loRaWANDeviceProfile_supportsClassC :: Lens' LoRaWANDeviceProfile (Maybe Bool)

The SupportsClassC value.

loRaWANDeviceProfile_pingSlotPeriod :: Lens' LoRaWANDeviceProfile (Maybe Natural)

The PingSlotPeriod value.