Network.AWS.IoTWireless.Types.LoRaWANUpdateGatewayTaskEntry

Description

Documentation

data LoRaWANUpdateGatewayTaskEntry

LoRaWANUpdateGatewayTaskEntry object.

See: newLoRaWANUpdateGatewayTaskEntry smart constructor.

Constructors

LoRaWANUpdateGatewayTaskEntry'

Fields

currentVersion :: Maybe LoRaWANGatewayVersion

The version of the gateways that should receive the update.

updateVersion :: Maybe LoRaWANGatewayVersion

The firmware version to update the gateway to.

newLoRaWANUpdateGatewayTaskEntry :: LoRaWANUpdateGatewayTaskEntry

Create a value of LoRaWANUpdateGatewayTaskEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currentVersion:LoRaWANUpdateGatewayTaskEntry', loRaWANUpdateGatewayTaskEntry_currentVersion - The version of the gateways that should receive the update.

$sel:updateVersion:LoRaWANUpdateGatewayTaskEntry', loRaWANUpdateGatewayTaskEntry_updateVersion - The firmware version to update the gateway to.

loRaWANUpdateGatewayTaskEntry_currentVersion :: Lens' LoRaWANUpdateGatewayTaskEntry (Maybe LoRaWANGatewayVersion)

The version of the gateways that should receive the update.

loRaWANUpdateGatewayTaskEntry_updateVersion :: Lens' LoRaWANUpdateGatewayTaskEntry (Maybe LoRaWANGatewayVersion)

The firmware version to update the gateway to.