Network.AWS.IoTWireless.Types.UpdateWirelessGatewayTaskCreate

Description

Documentation

data UpdateWirelessGatewayTaskCreate

UpdateWirelessGatewayTaskCreate object.

See: newUpdateWirelessGatewayTaskCreate smart constructor.

Constructors

UpdateWirelessGatewayTaskCreate'

Fields

updateDataSource :: Maybe Text

The link to the S3 bucket.

updateDataRole :: Maybe Text

The IAM role used to read data from the S3 bucket.

loRaWAN :: Maybe LoRaWANUpdateGatewayTaskCreate

The properties that relate to the LoRaWAN wireless gateway.

newUpdateWirelessGatewayTaskCreate :: UpdateWirelessGatewayTaskCreate

Create a value of UpdateWirelessGatewayTaskCreate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateDataSource:UpdateWirelessGatewayTaskCreate', updateWirelessGatewayTaskCreate_updateDataSource - The link to the S3 bucket.

$sel:updateDataRole:UpdateWirelessGatewayTaskCreate', updateWirelessGatewayTaskCreate_updateDataRole - The IAM role used to read data from the S3 bucket.

$sel:loRaWAN:UpdateWirelessGatewayTaskCreate', updateWirelessGatewayTaskCreate_loRaWAN - The properties that relate to the LoRaWAN wireless gateway.

updateWirelessGatewayTaskCreate_updateDataSource :: Lens' UpdateWirelessGatewayTaskCreate (Maybe Text)

The link to the S3 bucket.

updateWirelessGatewayTaskCreate_updateDataRole :: Lens' UpdateWirelessGatewayTaskCreate (Maybe Text)

The IAM role used to read data from the S3 bucket.

updateWirelessGatewayTaskCreate_loRaWAN :: Lens' UpdateWirelessGatewayTaskCreate (Maybe LoRaWANUpdateGatewayTaskCreate)

The properties that relate to the LoRaWAN wireless gateway.