Network.AWS.IoTWireless.Types.WirelessDeviceStatistics

Description

Documentation

data WirelessDeviceStatistics

Information about a wireless device's operation.

See: newWirelessDeviceStatistics smart constructor.

Constructors

WirelessDeviceStatistics'

Fields

sidewalk :: Maybe SidewalkListDevice

The Sidewalk account credentials.

arn :: Maybe Text

The Amazon Resource Name of the resource.

loRaWAN :: Maybe LoRaWANListDevice

LoRaWAN device info.

name :: Maybe Text

The name of the resource.

id :: Maybe Text

The ID of the wireless device reporting the data.

lastUplinkReceivedAt :: Maybe Text

The date and time when the most recent uplink was received.

type' :: Maybe WirelessDeviceType

The wireless device type.

destinationName :: Maybe Text

The name of the destination to which the device is assigned.

newWirelessDeviceStatistics :: WirelessDeviceStatistics

Create a value of WirelessDeviceStatistics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sidewalk:WirelessDeviceStatistics', wirelessDeviceStatistics_sidewalk - The Sidewalk account credentials.

$sel:arn:WirelessDeviceStatistics', wirelessDeviceStatistics_arn - The Amazon Resource Name of the resource.

$sel:loRaWAN:WirelessDeviceStatistics', wirelessDeviceStatistics_loRaWAN - LoRaWAN device info.

$sel:name:WirelessDeviceStatistics', wirelessDeviceStatistics_name - The name of the resource.

$sel:id:WirelessDeviceStatistics', wirelessDeviceStatistics_id - The ID of the wireless device reporting the data.

$sel:lastUplinkReceivedAt:WirelessDeviceStatistics', wirelessDeviceStatistics_lastUplinkReceivedAt - The date and time when the most recent uplink was received.

$sel:type':WirelessDeviceStatistics', wirelessDeviceStatistics_type - The wireless device type.

$sel:destinationName:WirelessDeviceStatistics', wirelessDeviceStatistics_destinationName - The name of the destination to which the device is assigned.

wirelessDeviceStatistics_sidewalk :: Lens' WirelessDeviceStatistics (Maybe SidewalkListDevice)

The Sidewalk account credentials.

wirelessDeviceStatistics_arn :: Lens' WirelessDeviceStatistics (Maybe Text)

The Amazon Resource Name of the resource.

wirelessDeviceStatistics_loRaWAN :: Lens' WirelessDeviceStatistics (Maybe LoRaWANListDevice)

LoRaWAN device info.

wirelessDeviceStatistics_name :: Lens' WirelessDeviceStatistics (Maybe Text)

The name of the resource.

wirelessDeviceStatistics_id :: Lens' WirelessDeviceStatistics (Maybe Text)

The ID of the wireless device reporting the data.

wirelessDeviceStatistics_lastUplinkReceivedAt :: Lens' WirelessDeviceStatistics (Maybe Text)

The date and time when the most recent uplink was received.

wirelessDeviceStatistics_type :: Lens' WirelessDeviceStatistics (Maybe WirelessDeviceType)

The wireless device type.

wirelessDeviceStatistics_destinationName :: Lens' WirelessDeviceStatistics (Maybe Text)

The name of the destination to which the device is assigned.