Network.AWS.IoTWireless.UpdateDestination

Description

Updates properties of a destination.

Creating a Request

data UpdateDestination

See: newUpdateDestination smart constructor.

Constructors

UpdateDestination'

Fields

expressionType :: Maybe ExpressionType

The type of value in Expression.

expression :: Maybe Text

The new rule name or topic rule to send messages to.

description :: Maybe Text

A new description of the resource.

roleArn :: Maybe Text

The ARN of the IAM Role that authorizes the destination.

name :: Text

The new name of the resource.

newUpdateDestination

Arguments

:: Text
$sel:name:UpdateDestination'

-> UpdateDestination

Create a value of UpdateDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expressionType:UpdateDestination', updateDestination_expressionType - The type of value in Expression.

$sel:expression:UpdateDestination', updateDestination_expression - The new rule name or topic rule to send messages to.

$sel:description:UpdateDestination', updateDestination_description - A new description of the resource.

$sel:roleArn:UpdateDestination', updateDestination_roleArn - The ARN of the IAM Role that authorizes the destination.

$sel:name:UpdateDestination', updateDestination_name - The new name of the resource.

Request Lenses

updateDestination_expressionType :: Lens' UpdateDestination (Maybe ExpressionType)

The type of value in Expression.

updateDestination_expression :: Lens' UpdateDestination (Maybe Text)

The new rule name or topic rule to send messages to.

updateDestination_description :: Lens' UpdateDestination (Maybe Text)

A new description of the resource.

updateDestination_roleArn :: Lens' UpdateDestination (Maybe Text)

The ARN of the IAM Role that authorizes the destination.

updateDestination_name :: Lens' UpdateDestination Text

The new name of the resource.

Destructuring the Response

data UpdateDestinationResponse

See: newUpdateDestinationResponse smart constructor.

Constructors

UpdateDestinationResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateDestinationResponse

Arguments

:: Int
$sel:httpStatus:UpdateDestinationResponse'

-> UpdateDestinationResponse

Create a value of UpdateDestinationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateDestinationResponse', updateDestinationResponse_httpStatus - The response's http status code.

Response Lenses

updateDestinationResponse_httpStatus :: Lens' UpdateDestinationResponse Int

The response's http status code.