Network.AWS.KMS.CreateCustomKeyStore

Description

Creates a custom key store that is associated with an CloudHSM cluster that you own and manage.

This operation is part of the Custom Key Store feature feature in KMS, which combines the convenience and extensive integration of KMS with the isolation and control of a single-tenant key store.

Before you create the custom key store, you must assemble the required elements, including an CloudHSM cluster that fulfills the requirements for a custom key store. For details about the required elements, see Assemble the Prerequisites in the Key Management Service Developer Guide.

When the operation completes successfully, it returns the ID of the new custom key store. Before you can use your new custom key store, you need to use the ConnectCustomKeyStore operation to connect the new key store to its CloudHSM cluster. Even if you are not going to use your custom key store immediately, you might want to connect it to verify that all settings are correct and then disconnect it until you are ready to use it.

For help with failures, see Troubleshooting a Custom Key Store in the Key Management Service Developer Guide.

Cross-account use: No. You cannot perform this operation on a custom key store in a different Amazon Web Services account.

Required permissions: kms:CreateCustomKeyStore (IAM policy).

Related operations:

ConnectCustomKeyStore

DeleteCustomKeyStore

DescribeCustomKeyStores

DisconnectCustomKeyStore

UpdateCustomKeyStore

Creating a Request

data CreateCustomKeyStore

See: newCreateCustomKeyStore smart constructor.

Constructors

CreateCustomKeyStore'

Fields

customKeyStoreName :: Text

Specifies a friendly name for the custom key store. The name must be unique in your Amazon Web Services account.

cloudHsmClusterId :: Text

Identifies the CloudHSM cluster for the custom key store. Enter the cluster ID of any active CloudHSM cluster that is not already associated with a custom key store. To find the cluster ID, use the DescribeClusters operation.

trustAnchorCertificate :: Text

Enter the content of the trust anchor certificate for the cluster. This is the content of the customerCA.crt file that you created when you initialized the cluster.

keyStorePassword :: Sensitive Text

Enter the password of the kmsuser crypto user (CU) account in the specified CloudHSM cluster. KMS logs into the cluster as this user to manage key material on your behalf.

The password must be a string of 7 to 32 characters. Its value is case sensitive.

This parameter tells KMS the kmsuser account password; it does not change the password in the CloudHSM cluster.

newCreateCustomKeyStore

Arguments

:: Text
$sel:customKeyStoreName:CreateCustomKeyStore'

-> Text
$sel:cloudHsmClusterId:CreateCustomKeyStore'

-> Text
$sel:trustAnchorCertificate:CreateCustomKeyStore'

-> Text
$sel:keyStorePassword:CreateCustomKeyStore'

-> CreateCustomKeyStore

Create a value of CreateCustomKeyStore with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customKeyStoreName:CreateCustomKeyStore', createCustomKeyStore_customKeyStoreName - Specifies a friendly name for the custom key store. The name must be unique in your Amazon Web Services account.

$sel:cloudHsmClusterId:CreateCustomKeyStore', createCustomKeyStore_cloudHsmClusterId - Identifies the CloudHSM cluster for the custom key store. Enter the cluster ID of any active CloudHSM cluster that is not already associated with a custom key store. To find the cluster ID, use the DescribeClusters operation.

$sel:trustAnchorCertificate:CreateCustomKeyStore', createCustomKeyStore_trustAnchorCertificate - Enter the content of the trust anchor certificate for the cluster. This is the content of the customerCA.crt file that you created when you initialized the cluster.

$sel:keyStorePassword:CreateCustomKeyStore', createCustomKeyStore_keyStorePassword - Enter the password of the kmsuser crypto user (CU) account in the specified CloudHSM cluster. KMS logs into the cluster as this user to manage key material on your behalf.

The password must be a string of 7 to 32 characters. Its value is case sensitive.

This parameter tells KMS the kmsuser account password; it does not change the password in the CloudHSM cluster.

Request Lenses

createCustomKeyStore_customKeyStoreName :: Lens' CreateCustomKeyStore Text

Specifies a friendly name for the custom key store. The name must be unique in your Amazon Web Services account.

createCustomKeyStore_cloudHsmClusterId :: Lens' CreateCustomKeyStore Text

Identifies the CloudHSM cluster for the custom key store. Enter the cluster ID of any active CloudHSM cluster that is not already associated with a custom key store. To find the cluster ID, use the DescribeClusters operation.

createCustomKeyStore_trustAnchorCertificate :: Lens' CreateCustomKeyStore Text

Enter the content of the trust anchor certificate for the cluster. This is the content of the customerCA.crt file that you created when you initialized the cluster.

createCustomKeyStore_keyStorePassword :: Lens' CreateCustomKeyStore Text

Enter the password of the kmsuser crypto user (CU) account in the specified CloudHSM cluster. KMS logs into the cluster as this user to manage key material on your behalf.

The password must be a string of 7 to 32 characters. Its value is case sensitive.

This parameter tells KMS the kmsuser account password; it does not change the password in the CloudHSM cluster.

Destructuring the Response

data CreateCustomKeyStoreResponse

See: newCreateCustomKeyStoreResponse smart constructor.

Constructors

CreateCustomKeyStoreResponse'

Fields

customKeyStoreId :: Maybe Text

A unique identifier for the new custom key store.

httpStatus :: Int

The response's http status code.

newCreateCustomKeyStoreResponse

Arguments

:: Int
$sel:httpStatus:CreateCustomKeyStoreResponse'

-> CreateCustomKeyStoreResponse

Create a value of CreateCustomKeyStoreResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customKeyStoreId:CreateCustomKeyStoreResponse', createCustomKeyStoreResponse_customKeyStoreId - A unique identifier for the new custom key store.

$sel:httpStatus:CreateCustomKeyStoreResponse', createCustomKeyStoreResponse_httpStatus - The response's http status code.

Response Lenses

createCustomKeyStoreResponse_customKeyStoreId :: Lens' CreateCustomKeyStoreResponse (Maybe Text)

A unique identifier for the new custom key store.

createCustomKeyStoreResponse_httpStatus :: Lens' CreateCustomKeyStoreResponse Int

The response's http status code.