Network.AWS.KMS.DeleteCustomKeyStore

Description

Deletes a custom key store. This operation does not delete the CloudHSM cluster that is associated with the custom key store, or affect any users or keys in the cluster.

The custom key store that you delete cannot contain any KMS KMS keys. Before deleting the key store, verify that you will never need to use any of the KMS keys in the key store for any cryptographic operations. Then, use ScheduleKeyDeletion to delete the KMS keys from the key store. When the scheduled waiting period expires, the ScheduleKeyDeletion operation deletes the KMS keys. Then it makes a best effort to delete the key material from the associated cluster. However, you might need to manually delete the orphaned key material from the cluster and its backups.

After all KMS keys are deleted from KMS, use DisconnectCustomKeyStore to disconnect the key store from KMS. Then, you can delete the custom key store.

Instead of deleting the custom key store, consider using DisconnectCustomKeyStore to disconnect it from KMS. While the key store is disconnected, you cannot create or use the KMS keys in the key store. But, you do not need to delete KMS keys and you can reconnect a disconnected custom key store at any time.

If the operation succeeds, it returns a JSON object with no properties.

This operation is part of the Custom Key Store feature feature in KMS, which combines the convenience and extensive integration of KMS with the isolation and control of a single-tenant key store.

Cross-account use: No. You cannot perform this operation on a custom key store in a different Amazon Web Services account.

Required permissions: kms:DeleteCustomKeyStore (IAM policy)

Related operations:

ConnectCustomKeyStore

CreateCustomKeyStore

DescribeCustomKeyStores

DisconnectCustomKeyStore

UpdateCustomKeyStore

Creating a Request

data DeleteCustomKeyStore

See: newDeleteCustomKeyStore smart constructor.

Constructors

DeleteCustomKeyStore'

Fields

customKeyStoreId :: Text

Enter the ID of the custom key store you want to delete. To find the ID of a custom key store, use the DescribeCustomKeyStores operation.

newDeleteCustomKeyStore

Arguments

:: Text
$sel:customKeyStoreId:DeleteCustomKeyStore'

-> DeleteCustomKeyStore

Create a value of DeleteCustomKeyStore with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customKeyStoreId:DeleteCustomKeyStore', deleteCustomKeyStore_customKeyStoreId - Enter the ID of the custom key store you want to delete. To find the ID of a custom key store, use the DescribeCustomKeyStores operation.

Request Lenses

deleteCustomKeyStore_customKeyStoreId :: Lens' DeleteCustomKeyStore Text

Enter the ID of the custom key store you want to delete. To find the ID of a custom key store, use the DescribeCustomKeyStores operation.

Destructuring the Response

data DeleteCustomKeyStoreResponse

See: newDeleteCustomKeyStoreResponse smart constructor.

Constructors

DeleteCustomKeyStoreResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteCustomKeyStoreResponse

Arguments

:: Int
$sel:httpStatus:DeleteCustomKeyStoreResponse'

-> DeleteCustomKeyStoreResponse

Create a value of DeleteCustomKeyStoreResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteCustomKeyStoreResponse', deleteCustomKeyStoreResponse_httpStatus - The response's http status code.

Response Lenses

deleteCustomKeyStoreResponse_httpStatus :: Lens' DeleteCustomKeyStoreResponse Int

The response's http status code.