Network.AWS.KMS.DescribeCustomKeyStores

Description

Gets information about custom key stores in the account and Region.

This operation is part of the Custom Key Store feature feature in KMS, which combines the convenience and extensive integration of KMS with the isolation and control of a single-tenant key store.

By default, this operation returns information about all custom key stores in the account and Region. To get only information about a particular custom key store, use either the CustomKeyStoreName or CustomKeyStoreId parameter (but not both).

To determine whether the custom key store is connected to its CloudHSM cluster, use the ConnectionState element in the response. If an attempt to connect the custom key store failed, the ConnectionState value is FAILED and the ConnectionErrorCode element in the response indicates the cause of the failure. For help interpreting the ConnectionErrorCode, see CustomKeyStoresListEntry.

Custom key stores have a DISCONNECTED connection state if the key store has never been connected or you use the DisconnectCustomKeyStore operation to disconnect it. If your custom key store state is CONNECTED but you are having trouble using it, make sure that its associated CloudHSM cluster is active and contains the minimum number of HSMs required for the operation, if any.

For help repairing your custom key store, see the Troubleshooting Custom Key Stores topic in the Key Management Service Developer Guide.

Cross-account use: No. You cannot perform this operation on a custom key store in a different Amazon Web Services account.

Required permissions: kms:DescribeCustomKeyStores (IAM policy)

Related operations:

ConnectCustomKeyStore

CreateCustomKeyStore

DeleteCustomKeyStore

DisconnectCustomKeyStore

UpdateCustomKeyStore

Creating a Request

data DescribeCustomKeyStores

See: newDescribeCustomKeyStores smart constructor.

Constructors

DescribeCustomKeyStores'

Fields

customKeyStoreName :: Maybe Text

Gets only information about the specified custom key store. Enter the friendly name of the custom key store.

By default, this operation gets information about all custom key stores in the account and Region. To limit the output to a particular custom key store, you can use either the CustomKeyStoreId or CustomKeyStoreName parameter, but not both.

customKeyStoreId :: Maybe Text

Gets only information about the specified custom key store. Enter the key store ID.

By default, this operation gets information about all custom key stores in the account and Region. To limit the output to a particular custom key store, you can use either the CustomKeyStoreId or CustomKeyStoreName parameter, but not both.

limit :: Maybe Natural

Use this parameter to specify the maximum number of items to return. When this value is present, KMS does not return more than the specified number of items, but it might return fewer.

marker :: Maybe Text

Use this parameter in a subsequent request after you receive a response with truncated results. Set it to the value of NextMarker from the truncated response you just received.

newDescribeCustomKeyStores :: DescribeCustomKeyStores

Create a value of DescribeCustomKeyStores with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customKeyStoreName:DescribeCustomKeyStores', describeCustomKeyStores_customKeyStoreName - Gets only information about the specified custom key store. Enter the friendly name of the custom key store.

By default, this operation gets information about all custom key stores in the account and Region. To limit the output to a particular custom key store, you can use either the CustomKeyStoreId or CustomKeyStoreName parameter, but not both.

$sel:customKeyStoreId:DescribeCustomKeyStores', describeCustomKeyStores_customKeyStoreId - Gets only information about the specified custom key store. Enter the key store ID.

By default, this operation gets information about all custom key stores in the account and Region. To limit the output to a particular custom key store, you can use either the CustomKeyStoreId or CustomKeyStoreName parameter, but not both.

$sel:limit:DescribeCustomKeyStores', describeCustomKeyStores_limit - Use this parameter to specify the maximum number of items to return. When this value is present, KMS does not return more than the specified number of items, but it might return fewer.

$sel:marker:DescribeCustomKeyStores', describeCustomKeyStores_marker - Use this parameter in a subsequent request after you receive a response with truncated results. Set it to the value of NextMarker from the truncated response you just received.

Request Lenses

describeCustomKeyStores_customKeyStoreName :: Lens' DescribeCustomKeyStores (Maybe Text)

Gets only information about the specified custom key store. Enter the friendly name of the custom key store.

By default, this operation gets information about all custom key stores in the account and Region. To limit the output to a particular custom key store, you can use either the CustomKeyStoreId or CustomKeyStoreName parameter, but not both.

describeCustomKeyStores_customKeyStoreId :: Lens' DescribeCustomKeyStores (Maybe Text)

Gets only information about the specified custom key store. Enter the key store ID.

By default, this operation gets information about all custom key stores in the account and Region. To limit the output to a particular custom key store, you can use either the CustomKeyStoreId or CustomKeyStoreName parameter, but not both.

describeCustomKeyStores_limit :: Lens' DescribeCustomKeyStores (Maybe Natural)

Use this parameter to specify the maximum number of items to return. When this value is present, KMS does not return more than the specified number of items, but it might return fewer.

describeCustomKeyStores_marker :: Lens' DescribeCustomKeyStores (Maybe Text)

Use this parameter in a subsequent request after you receive a response with truncated results. Set it to the value of NextMarker from the truncated response you just received.

Destructuring the Response

data DescribeCustomKeyStoresResponse

See: newDescribeCustomKeyStoresResponse smart constructor.

Constructors

DescribeCustomKeyStoresResponse'

Fields

customKeyStores :: Maybe [CustomKeyStoresListEntry]

Contains metadata about each custom key store.

nextMarker :: Maybe Text

When Truncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent request.

truncated :: Maybe Bool

A flag that indicates whether there are more items in the list. When this value is true, the list in this response is truncated. To get more items, pass the value of the NextMarker element in thisresponse to the Marker parameter in a subsequent request.

httpStatus :: Int

The response's http status code.

newDescribeCustomKeyStoresResponse

Arguments

:: Int
$sel:httpStatus:DescribeCustomKeyStoresResponse'

-> DescribeCustomKeyStoresResponse

Create a value of DescribeCustomKeyStoresResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customKeyStores:DescribeCustomKeyStoresResponse', describeCustomKeyStoresResponse_customKeyStores - Contains metadata about each custom key store.

$sel:nextMarker:DescribeCustomKeyStoresResponse', describeCustomKeyStoresResponse_nextMarker - When Truncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent request.

$sel:truncated:DescribeCustomKeyStoresResponse', describeCustomKeyStoresResponse_truncated - A flag that indicates whether there are more items in the list. When this value is true, the list in this response is truncated. To get more items, pass the value of the NextMarker element in thisresponse to the Marker parameter in a subsequent request.

$sel:httpStatus:DescribeCustomKeyStoresResponse', describeCustomKeyStoresResponse_httpStatus - The response's http status code.

Response Lenses

describeCustomKeyStoresResponse_customKeyStores :: Lens' DescribeCustomKeyStoresResponse (Maybe [CustomKeyStoresListEntry])

Contains metadata about each custom key store.

describeCustomKeyStoresResponse_nextMarker :: Lens' DescribeCustomKeyStoresResponse (Maybe Text)

When Truncated is true, this element is present and contains the value to use for the Marker parameter in a subsequent request.

describeCustomKeyStoresResponse_truncated :: Lens' DescribeCustomKeyStoresResponse (Maybe Bool)

A flag that indicates whether there are more items in the list. When this value is true, the list in this response is truncated. To get more items, pass the value of the NextMarker element in thisresponse to the Marker parameter in a subsequent request.

describeCustomKeyStoresResponse_httpStatus :: Lens' DescribeCustomKeyStoresResponse Int

The response's http status code.