Network.AWS.KMS.EnableKey

Description

Sets the key state of a KMS key to enabled. This allows you to use the KMS key for cryptographic operations.

The KMS key that you use for this operation must be in a compatible key state. For details, see Key state: Effect on your KMS key in the Key Management Service Developer Guide.

Cross-account use: No. You cannot perform this operation on a KMS key in a different Amazon Web Services account.

Required permissions: kms:EnableKey (key policy)

Related operations: DisableKey

Creating a Request

data EnableKey

See: newEnableKey smart constructor.

Constructors

EnableKey'

Fields

keyId :: Text

Identifies the KMS key to enable.

Specify the key ID or key ARN of the KMS key.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

To get the key ID and key ARN for a KMS key, use ListKeys or DescribeKey.

newEnableKey

Arguments

:: Text
$sel:keyId:EnableKey'

-> EnableKey

Create a value of EnableKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyId:EnableKey', enableKey_keyId - Identifies the KMS key to enable.

Specify the key ID or key ARN of the KMS key.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

To get the key ID and key ARN for a KMS key, use ListKeys or DescribeKey.

Request Lenses

enableKey_keyId :: Lens' EnableKey Text

Identifies the KMS key to enable.

Specify the key ID or key ARN of the KMS key.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

To get the key ID and key ARN for a KMS key, use ListKeys or DescribeKey.

Destructuring the Response

data EnableKeyResponse

See: newEnableKeyResponse smart constructor.

Constructors

EnableKeyResponse'

newEnableKeyResponse :: EnableKeyResponse

Create a value of EnableKeyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.