Network.AWS.KMS.GenerateDataKeyPairWithoutPlaintext

Description

Generates a unique asymmetric data key pair. The GenerateDataKeyPairWithoutPlaintext operation returns a plaintext public key and a copy of the private key that is encrypted under the symmetric KMS key you specify. Unlike GenerateDataKeyPair, this operation does not return a plaintext private key.

You can use the public key that GenerateDataKeyPairWithoutPlaintext returns to encrypt data or verify a signature outside of KMS. Then, store the encrypted private key with the data. When you are ready to decrypt data or sign a message, you can use the Decrypt operation to decrypt the encrypted private key.

To generate a data key pair, you must specify a symmetric KMS key to encrypt the private key in a data key pair. You cannot use an asymmetric KMS key or a KMS key in a custom key store. To get the type and origin of your KMS key, use the DescribeKey operation.

Use the KeyPairSpec parameter to choose an RSA or Elliptic Curve (ECC) data key pair. KMS recommends that your use ECC key pairs for signing, and use RSA key pairs for either encryption or signing, but not both. However, KMS cannot enforce any restrictions on the use of data key pairs outside of KMS.

GenerateDataKeyPairWithoutPlaintext returns a unique data key pair for each request. The bytes in the key are not related to the caller or KMS key that is used to encrypt the private key. The public key is a DER-encoded X.509 SubjectPublicKeyInfo, as specified in RFC 5280.

You can use the optional encryption context to add additional security to the encryption operation. If you specify an EncryptionContext, you must specify the same encryption context (a case-sensitive exact match) when decrypting the encrypted data key. Otherwise, the request to decrypt fails with an InvalidCiphertextException. For more information, see Encryption Context in the Key Management Service Developer Guide.

The KMS key that you use for this operation must be in a compatible key state. For details, see Key state: Effect on your KMS key in the Key Management Service Developer Guide.

Cross-account use: Yes. To perform this operation with a KMS key in a different Amazon Web Services account, specify the key ARN or alias ARN in the value of the KeyId parameter.

Required permissions: kms:GenerateDataKeyPairWithoutPlaintext (key policy)

Related operations:

Decrypt

Encrypt

GenerateDataKey

GenerateDataKeyPair

GenerateDataKeyWithoutPlaintext

Creating a Request

data GenerateDataKeyPairWithoutPlaintext

See: newGenerateDataKeyPairWithoutPlaintext smart constructor.

Constructors

GenerateDataKeyPairWithoutPlaintext'

Fields

encryptionContext :: Maybe (HashMap Text Text)

Specifies the encryption context that will be used when encrypting the private key in the data key pair.

An encryption context is a collection of non-secret key-value pairs that represents additional authenticated data. When you use an encryption context to encrypt data, you must specify the same (an exact case-sensitive match) encryption context to decrypt the data. An encryption context is optional when encrypting with a symmetric KMS key, but it is highly recommended.

For more information, see Encryption Context in the Key Management Service Developer Guide.

grantTokens :: Maybe [Text]

A list of grant tokens.

Use a grant token when your permission to call this operation comes from a new grant that has not yet achieved eventual consistency. For more information, see Grant token and Using a grant token in the Key Management Service Developer Guide.

keyId :: Text

Specifies the KMS key that encrypts the private key in the data key pair. You must specify a symmetric KMS key. You cannot use an asymmetric KMS key or a KMS key in a custom key store. To get the type and origin of your KMS key, use the DescribeKey operation.

To specify a KMS key, use its key ID, key ARN, alias name, or alias ARN. When using an alias name, prefix it with "alias/". To specify a KMS key in a different Amazon Web Services account, you must use the key ARN or alias ARN.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Alias name: alias/ExampleAlias

Alias ARN: arn:aws:kms:us-east-2:111122223333:alias/ExampleAlias

To get the key ID and key ARN for a KMS key, use ListKeys or DescribeKey. To get the alias name and alias ARN, use ListAliases.

keyPairSpec :: DataKeyPairSpec

Determines the type of data key pair that is generated.

The KMS rule that restricts the use of asymmetric RSA KMS keys to encrypt and decrypt or to sign and verify (but not both), and the rule that permits you to use ECC KMS keys only to sign and verify, are not effective on data key pairs, which are used outside of KMS.

newGenerateDataKeyPairWithoutPlaintext

Arguments

:: Text
$sel:keyId:GenerateDataKeyPairWithoutPlaintext'

-> DataKeyPairSpec
$sel:keyPairSpec:GenerateDataKeyPairWithoutPlaintext'

-> GenerateDataKeyPairWithoutPlaintext

Create a value of GenerateDataKeyPairWithoutPlaintext with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encryptionContext:GenerateDataKeyPairWithoutPlaintext', generateDataKeyPairWithoutPlaintext_encryptionContext - Specifies the encryption context that will be used when encrypting the private key in the data key pair.

An encryption context is a collection of non-secret key-value pairs that represents additional authenticated data. When you use an encryption context to encrypt data, you must specify the same (an exact case-sensitive match) encryption context to decrypt the data. An encryption context is optional when encrypting with a symmetric KMS key, but it is highly recommended.

For more information, see Encryption Context in the Key Management Service Developer Guide.

$sel:grantTokens:GenerateDataKeyPairWithoutPlaintext', generateDataKeyPairWithoutPlaintext_grantTokens - A list of grant tokens.

Use a grant token when your permission to call this operation comes from a new grant that has not yet achieved eventual consistency. For more information, see Grant token and Using a grant token in the Key Management Service Developer Guide.

$sel:keyId:GenerateDataKeyPairWithoutPlaintext', generateDataKeyPairWithoutPlaintext_keyId - Specifies the KMS key that encrypts the private key in the data key pair. You must specify a symmetric KMS key. You cannot use an asymmetric KMS key or a KMS key in a custom key store. To get the type and origin of your KMS key, use the DescribeKey operation.

To specify a KMS key, use its key ID, key ARN, alias name, or alias ARN. When using an alias name, prefix it with "alias/". To specify a KMS key in a different Amazon Web Services account, you must use the key ARN or alias ARN.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Alias name: alias/ExampleAlias

Alias ARN: arn:aws:kms:us-east-2:111122223333:alias/ExampleAlias

To get the key ID and key ARN for a KMS key, use ListKeys or DescribeKey. To get the alias name and alias ARN, use ListAliases.

$sel:keyPairSpec:GenerateDataKeyPairWithoutPlaintext', generateDataKeyPairWithoutPlaintext_keyPairSpec - Determines the type of data key pair that is generated.

The KMS rule that restricts the use of asymmetric RSA KMS keys to encrypt and decrypt or to sign and verify (but not both), and the rule that permits you to use ECC KMS keys only to sign and verify, are not effective on data key pairs, which are used outside of KMS.

Request Lenses

generateDataKeyPairWithoutPlaintext_encryptionContext :: Lens' GenerateDataKeyPairWithoutPlaintext (Maybe (HashMap Text Text))

Specifies the encryption context that will be used when encrypting the private key in the data key pair.

An encryption context is a collection of non-secret key-value pairs that represents additional authenticated data. When you use an encryption context to encrypt data, you must specify the same (an exact case-sensitive match) encryption context to decrypt the data. An encryption context is optional when encrypting with a symmetric KMS key, but it is highly recommended.

For more information, see Encryption Context in the Key Management Service Developer Guide.

generateDataKeyPairWithoutPlaintext_grantTokens :: Lens' GenerateDataKeyPairWithoutPlaintext (Maybe [Text])

A list of grant tokens.

Use a grant token when your permission to call this operation comes from a new grant that has not yet achieved eventual consistency. For more information, see Grant token and Using a grant token in the Key Management Service Developer Guide.

generateDataKeyPairWithoutPlaintext_keyId :: Lens' GenerateDataKeyPairWithoutPlaintext Text

Specifies the KMS key that encrypts the private key in the data key pair. You must specify a symmetric KMS key. You cannot use an asymmetric KMS key or a KMS key in a custom key store. To get the type and origin of your KMS key, use the DescribeKey operation.

To specify a KMS key, use its key ID, key ARN, alias name, or alias ARN. When using an alias name, prefix it with "alias/". To specify a KMS key in a different Amazon Web Services account, you must use the key ARN or alias ARN.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Alias name: alias/ExampleAlias

Alias ARN: arn:aws:kms:us-east-2:111122223333:alias/ExampleAlias

To get the key ID and key ARN for a KMS key, use ListKeys or DescribeKey. To get the alias name and alias ARN, use ListAliases.

generateDataKeyPairWithoutPlaintext_keyPairSpec :: Lens' GenerateDataKeyPairWithoutPlaintext DataKeyPairSpec

Determines the type of data key pair that is generated.

The KMS rule that restricts the use of asymmetric RSA KMS keys to encrypt and decrypt or to sign and verify (but not both), and the rule that permits you to use ECC KMS keys only to sign and verify, are not effective on data key pairs, which are used outside of KMS.

Destructuring the Response

data GenerateDataKeyPairWithoutPlaintextResponse

See: newGenerateDataKeyPairWithoutPlaintextResponse smart constructor.

Constructors

GenerateDataKeyPairWithoutPlaintextResponse'

Fields

keyId :: Maybe Text

The Amazon Resource Name (key ARN) of the KMS key that encrypted the private key.

publicKey :: Maybe Base64

The public key (in plaintext).

keyPairSpec :: Maybe DataKeyPairSpec

The type of data key pair that was generated.

privateKeyCiphertextBlob :: Maybe Base64

The encrypted copy of the private key. When you use the HTTP API or the Amazon Web Services CLI, the value is Base64-encoded. Otherwise, it is not Base64-encoded.

httpStatus :: Int

The response's http status code.

newGenerateDataKeyPairWithoutPlaintextResponse

Arguments

:: Int
$sel:httpStatus:GenerateDataKeyPairWithoutPlaintextResponse'

-> GenerateDataKeyPairWithoutPlaintextResponse

Create a value of GenerateDataKeyPairWithoutPlaintextResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyId:GenerateDataKeyPairWithoutPlaintext', generateDataKeyPairWithoutPlaintextResponse_keyId - The Amazon Resource Name (key ARN) of the KMS key that encrypted the private key.

$sel:publicKey:GenerateDataKeyPairWithoutPlaintextResponse', generateDataKeyPairWithoutPlaintextResponse_publicKey - The public key (in plaintext).-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:keyPairSpec:GenerateDataKeyPairWithoutPlaintext', generateDataKeyPairWithoutPlaintextResponse_keyPairSpec - The type of data key pair that was generated.

$sel:privateKeyCiphertextBlob:GenerateDataKeyPairWithoutPlaintextResponse', generateDataKeyPairWithoutPlaintextResponse_privateKeyCiphertextBlob - The encrypted copy of the private key. When you use the HTTP API or the Amazon Web Services CLI, the value is Base64-encoded. Otherwise, it is not Base64-encoded.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:httpStatus:GenerateDataKeyPairWithoutPlaintextResponse', generateDataKeyPairWithoutPlaintextResponse_httpStatus - The response's http status code.

Response Lenses

generateDataKeyPairWithoutPlaintextResponse_keyId :: Lens' GenerateDataKeyPairWithoutPlaintextResponse (Maybe Text)

The Amazon Resource Name (key ARN) of the KMS key that encrypted the private key.

generateDataKeyPairWithoutPlaintextResponse_publicKey :: Lens' GenerateDataKeyPairWithoutPlaintextResponse (Maybe ByteString)

The public key (in plaintext).-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

generateDataKeyPairWithoutPlaintextResponse_keyPairSpec :: Lens' GenerateDataKeyPairWithoutPlaintextResponse (Maybe DataKeyPairSpec)

The type of data key pair that was generated.

generateDataKeyPairWithoutPlaintextResponse_privateKeyCiphertextBlob :: Lens' GenerateDataKeyPairWithoutPlaintextResponse (Maybe ByteString)

The encrypted copy of the private key. When you use the HTTP API or the Amazon Web Services CLI, the value is Base64-encoded. Otherwise, it is not Base64-encoded.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

generateDataKeyPairWithoutPlaintextResponse_httpStatus :: Lens' GenerateDataKeyPairWithoutPlaintextResponse Int

The response's http status code.