Network.AWS.KMS.GetKeyPolicy

Description

Gets a key policy attached to the specified KMS key.

Cross-account use: No. You cannot perform this operation on a KMS key in a different Amazon Web Services account.

Required permissions: kms:GetKeyPolicy (key policy)

Related operations: PutKeyPolicy

Creating a Request

data GetKeyPolicy

See: newGetKeyPolicy smart constructor.

Constructors

GetKeyPolicy'

Fields

keyId :: Text

Gets the key policy for the specified KMS key.

Specify the key ID or key ARN of the KMS key.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

To get the key ID and key ARN for a KMS key, use ListKeys or DescribeKey.

policyName :: Text

Specifies the name of the key policy. The only valid name is default. To get the names of key policies, use ListKeyPolicies.

newGetKeyPolicy

Arguments

:: Text
$sel:keyId:GetKeyPolicy'

-> Text
$sel:policyName:GetKeyPolicy'

-> GetKeyPolicy

Create a value of GetKeyPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyId:GetKeyPolicy', getKeyPolicy_keyId - Gets the key policy for the specified KMS key.

Specify the key ID or key ARN of the KMS key.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

To get the key ID and key ARN for a KMS key, use ListKeys or DescribeKey.

$sel:policyName:GetKeyPolicy', getKeyPolicy_policyName - Specifies the name of the key policy. The only valid name is default. To get the names of key policies, use ListKeyPolicies.

Request Lenses

getKeyPolicy_keyId :: Lens' GetKeyPolicy Text

Gets the key policy for the specified KMS key.

Specify the key ID or key ARN of the KMS key.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

To get the key ID and key ARN for a KMS key, use ListKeys or DescribeKey.

getKeyPolicy_policyName :: Lens' GetKeyPolicy Text

Specifies the name of the key policy. The only valid name is default. To get the names of key policies, use ListKeyPolicies.

Destructuring the Response

data GetKeyPolicyResponse

See: newGetKeyPolicyResponse smart constructor.

Constructors

GetKeyPolicyResponse'

Fields

policy :: Maybe Text

A key policy document in JSON format.

httpStatus :: Int

The response's http status code.

newGetKeyPolicyResponse

Arguments

:: Int
$sel:httpStatus:GetKeyPolicyResponse'

-> GetKeyPolicyResponse

Create a value of GetKeyPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policy:GetKeyPolicyResponse', getKeyPolicyResponse_policy - A key policy document in JSON format.

$sel:httpStatus:GetKeyPolicyResponse', getKeyPolicyResponse_httpStatus - The response's http status code.

Response Lenses

getKeyPolicyResponse_policy :: Lens' GetKeyPolicyResponse (Maybe Text)

A key policy document in JSON format.

getKeyPolicyResponse_httpStatus :: Lens' GetKeyPolicyResponse Int

The response's http status code.