Network.AWS.KMS.Types.AliasListEntry

Description

Documentation

data AliasListEntry

Contains information about an alias.

See: newAliasListEntry smart constructor.

Constructors

AliasListEntry'

Fields

lastUpdatedDate :: Maybe POSIX

Date and time that the alias was most recently associated with a KMS key in the account and Region. Formatted as Unix time.

creationDate :: Maybe POSIX

Date and time that the alias was most recently created in the account and Region. Formatted as Unix time.

aliasName :: Maybe Text

String that contains the alias. This value begins with alias/.

targetKeyId :: Maybe Text

String that contains the key identifier of the KMS key associated with the alias.

aliasArn :: Maybe Text

String that contains the key ARN.

newAliasListEntry :: AliasListEntry

Create a value of AliasListEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedDate:AliasListEntry', aliasListEntry_lastUpdatedDate - Date and time that the alias was most recently associated with a KMS key in the account and Region. Formatted as Unix time.

$sel:creationDate:AliasListEntry', aliasListEntry_creationDate - Date and time that the alias was most recently created in the account and Region. Formatted as Unix time.

$sel:aliasName:AliasListEntry', aliasListEntry_aliasName - String that contains the alias. This value begins with alias/.

$sel:targetKeyId:AliasListEntry', aliasListEntry_targetKeyId - String that contains the key identifier of the KMS key associated with the alias.

$sel:aliasArn:AliasListEntry', aliasListEntry_aliasArn - String that contains the key ARN.

aliasListEntry_lastUpdatedDate :: Lens' AliasListEntry (Maybe UTCTime)

Date and time that the alias was most recently associated with a KMS key in the account and Region. Formatted as Unix time.

aliasListEntry_creationDate :: Lens' AliasListEntry (Maybe UTCTime)

Date and time that the alias was most recently created in the account and Region. Formatted as Unix time.

aliasListEntry_aliasName :: Lens' AliasListEntry (Maybe Text)

String that contains the alias. This value begins with alias/.

aliasListEntry_targetKeyId :: Lens' AliasListEntry (Maybe Text)

String that contains the key identifier of the KMS key associated with the alias.

aliasListEntry_aliasArn :: Lens' AliasListEntry (Maybe Text)

String that contains the key ARN.