Network.AWS.KMS.Types.GrantListEntry

Description

Documentation

data GrantListEntry

Contains information about a grant.

See: newGrantListEntry smart constructor.

Constructors

GrantListEntry'

Fields

constraints :: Maybe GrantConstraints

A list of key-value pairs that must be present in the encryption context of certain subsequent operations that the grant allows.

operations :: Maybe [GrantOperation]

The list of operations permitted by the grant.

creationDate :: Maybe POSIX

The date and time when the grant was created.

name :: Maybe Text

The friendly name that identifies the grant. If a name was provided in the CreateGrant request, that name is returned. Otherwise this value is null.

granteePrincipal :: Maybe Text

The identity that gets the permissions in the grant.

The GranteePrincipal field in the ListGrants response usually contains the user or role designated as the grantee principal in the grant. However, when the grantee principal in the grant is an Amazon Web Services service, the GranteePrincipal field contains the service principal, which might represent several different grantee principals.

grantId :: Maybe Text

The unique identifier for the grant.

issuingAccount :: Maybe Text

The Amazon Web Services account under which the grant was issued.

retiringPrincipal :: Maybe Text

The principal that can retire the grant.

keyId :: Maybe Text

The unique identifier for the KMS key to which the grant applies.

newGrantListEntry :: GrantListEntry

Create a value of GrantListEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:constraints:GrantListEntry', grantListEntry_constraints - A list of key-value pairs that must be present in the encryption context of certain subsequent operations that the grant allows.

$sel:operations:GrantListEntry', grantListEntry_operations - The list of operations permitted by the grant.

$sel:creationDate:GrantListEntry', grantListEntry_creationDate - The date and time when the grant was created.

$sel:name:GrantListEntry', grantListEntry_name - The friendly name that identifies the grant. If a name was provided in the CreateGrant request, that name is returned. Otherwise this value is null.

$sel:granteePrincipal:GrantListEntry', grantListEntry_granteePrincipal - The identity that gets the permissions in the grant.

The GranteePrincipal field in the ListGrants response usually contains the user or role designated as the grantee principal in the grant. However, when the grantee principal in the grant is an Amazon Web Services service, the GranteePrincipal field contains the service principal, which might represent several different grantee principals.

$sel:grantId:GrantListEntry', grantListEntry_grantId - The unique identifier for the grant.

$sel:issuingAccount:GrantListEntry', grantListEntry_issuingAccount - The Amazon Web Services account under which the grant was issued.

$sel:retiringPrincipal:GrantListEntry', grantListEntry_retiringPrincipal - The principal that can retire the grant.

$sel:keyId:GrantListEntry', grantListEntry_keyId - The unique identifier for the KMS key to which the grant applies.

grantListEntry_constraints :: Lens' GrantListEntry (Maybe GrantConstraints)

A list of key-value pairs that must be present in the encryption context of certain subsequent operations that the grant allows.

grantListEntry_operations :: Lens' GrantListEntry (Maybe [GrantOperation])

The list of operations permitted by the grant.

grantListEntry_creationDate :: Lens' GrantListEntry (Maybe UTCTime)

The date and time when the grant was created.

grantListEntry_name :: Lens' GrantListEntry (Maybe Text)

The friendly name that identifies the grant. If a name was provided in the CreateGrant request, that name is returned. Otherwise this value is null.

grantListEntry_granteePrincipal :: Lens' GrantListEntry (Maybe Text)

The identity that gets the permissions in the grant.

The GranteePrincipal field in the ListGrants response usually contains the user or role designated as the grantee principal in the grant. However, when the grantee principal in the grant is an Amazon Web Services service, the GranteePrincipal field contains the service principal, which might represent several different grantee principals.

grantListEntry_grantId :: Lens' GrantListEntry (Maybe Text)

The unique identifier for the grant.

grantListEntry_issuingAccount :: Lens' GrantListEntry (Maybe Text)

The Amazon Web Services account under which the grant was issued.

grantListEntry_retiringPrincipal :: Lens' GrantListEntry (Maybe Text)

The principal that can retire the grant.

grantListEntry_keyId :: Lens' GrantListEntry (Maybe Text)

The unique identifier for the KMS key to which the grant applies.