Network.AWS.Kafka.BatchAssociateScramSecret

Description

Associates one or more Scram Secrets with an Amazon MSK cluster.

Creating a Request

data BatchAssociateScramSecret

Associates sasl scram secrets to cluster.

See: newBatchAssociateScramSecret smart constructor.

Constructors

BatchAssociateScramSecret'

Fields

clusterArn :: Text

The Amazon Resource Name (ARN) of the cluster to be updated.

secretArnList :: [Text]

List of AWS Secrets Manager secret ARNs.

newBatchAssociateScramSecret

Arguments

:: Text
$sel:clusterArn:BatchAssociateScramSecret'

-> BatchAssociateScramSecret

Create a value of BatchAssociateScramSecret with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterArn:BatchAssociateScramSecret', batchAssociateScramSecret_clusterArn - The Amazon Resource Name (ARN) of the cluster to be updated.

$sel:secretArnList:BatchAssociateScramSecret', batchAssociateScramSecret_secretArnList - List of AWS Secrets Manager secret ARNs.

Request Lenses

batchAssociateScramSecret_clusterArn :: Lens' BatchAssociateScramSecret Text

The Amazon Resource Name (ARN) of the cluster to be updated.

batchAssociateScramSecret_secretArnList :: Lens' BatchAssociateScramSecret [Text]

List of AWS Secrets Manager secret ARNs.

Destructuring the Response

data BatchAssociateScramSecretResponse

See: newBatchAssociateScramSecretResponse smart constructor.

Constructors

BatchAssociateScramSecretResponse'

Fields

clusterArn :: Maybe Text

The Amazon Resource Name (ARN) of the cluster.

unprocessedScramSecrets :: Maybe [UnprocessedScramSecret]

List of errors when associating secrets to cluster.

httpStatus :: Int

The response's http status code.

newBatchAssociateScramSecretResponse

Arguments

:: Int
$sel:httpStatus:BatchAssociateScramSecretResponse'

-> BatchAssociateScramSecretResponse

Create a value of BatchAssociateScramSecretResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterArn:BatchAssociateScramSecret', batchAssociateScramSecretResponse_clusterArn - The Amazon Resource Name (ARN) of the cluster.

$sel:unprocessedScramSecrets:BatchAssociateScramSecretResponse', batchAssociateScramSecretResponse_unprocessedScramSecrets - List of errors when associating secrets to cluster.

$sel:httpStatus:BatchAssociateScramSecretResponse', batchAssociateScramSecretResponse_httpStatus - The response's http status code.

Response Lenses

batchAssociateScramSecretResponse_clusterArn :: Lens' BatchAssociateScramSecretResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

batchAssociateScramSecretResponse_unprocessedScramSecrets :: Lens' BatchAssociateScramSecretResponse (Maybe [UnprocessedScramSecret])

List of errors when associating secrets to cluster.

batchAssociateScramSecretResponse_httpStatus :: Lens' BatchAssociateScramSecretResponse Int

The response's http status code.