Network.AWS.Kafka.CreateCluster

Description

Creates a new MSK cluster.

Creating a Request

data CreateCluster

See: newCreateCluster smart constructor.

Constructors

CreateCluster'

Fields

enhancedMonitoring :: Maybe EnhancedMonitoring

Specifies the level of monitoring for the MSK cluster. The possible values are DEFAULT, PER_BROKER, PER_TOPIC_PER_BROKER, and PER_TOPIC_PER_PARTITION.

openMonitoring :: Maybe OpenMonitoringInfo

The settings for open monitoring.

configurationInfo :: Maybe ConfigurationInfo

Represents the configuration that you want MSK to use for the brokers in a cluster.

loggingInfo :: Maybe LoggingInfo

clientAuthentication :: Maybe ClientAuthentication

Includes all client authentication related information.

tags :: Maybe (HashMap Text Text)

Create tags when creating the cluster.

encryptionInfo :: Maybe EncryptionInfo

Includes all encryption-related information.

brokerNodeGroupInfo :: BrokerNodeGroupInfo

Information about the broker nodes in the cluster.

kafkaVersion :: Text

The version of Apache Kafka.

numberOfBrokerNodes :: Natural

The number of broker nodes in the cluster.

clusterName :: Text

The name of the cluster.

newCreateCluster

Arguments

:: BrokerNodeGroupInfo
$sel:brokerNodeGroupInfo:CreateCluster'

-> Text
$sel:kafkaVersion:CreateCluster'

-> Natural
$sel:numberOfBrokerNodes:CreateCluster'

-> Text
$sel:clusterName:CreateCluster'

-> CreateCluster

Create a value of CreateCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enhancedMonitoring:CreateCluster', createCluster_enhancedMonitoring - Specifies the level of monitoring for the MSK cluster. The possible values are DEFAULT, PER_BROKER, PER_TOPIC_PER_BROKER, and PER_TOPIC_PER_PARTITION.

$sel:openMonitoring:CreateCluster', createCluster_openMonitoring - The settings for open monitoring.

$sel:configurationInfo:CreateCluster', createCluster_configurationInfo - Represents the configuration that you want MSK to use for the brokers in a cluster.

$sel:loggingInfo:CreateCluster', createCluster_loggingInfo - Undocumented member.

$sel:clientAuthentication:CreateCluster', createCluster_clientAuthentication - Includes all client authentication related information.

$sel:tags:CreateCluster', createCluster_tags - Create tags when creating the cluster.

$sel:encryptionInfo:CreateCluster', createCluster_encryptionInfo - Includes all encryption-related information.

$sel:brokerNodeGroupInfo:CreateCluster', createCluster_brokerNodeGroupInfo - Information about the broker nodes in the cluster.

$sel:kafkaVersion:CreateCluster', createCluster_kafkaVersion - The version of Apache Kafka.

$sel:numberOfBrokerNodes:CreateCluster', createCluster_numberOfBrokerNodes - The number of broker nodes in the cluster.

$sel:clusterName:CreateCluster', createCluster_clusterName - The name of the cluster.

Request Lenses

createCluster_enhancedMonitoring :: Lens' CreateCluster (Maybe EnhancedMonitoring)

Specifies the level of monitoring for the MSK cluster. The possible values are DEFAULT, PER_BROKER, PER_TOPIC_PER_BROKER, and PER_TOPIC_PER_PARTITION.

createCluster_openMonitoring :: Lens' CreateCluster (Maybe OpenMonitoringInfo)

The settings for open monitoring.

createCluster_configurationInfo :: Lens' CreateCluster (Maybe ConfigurationInfo)

Represents the configuration that you want MSK to use for the brokers in a cluster.

createCluster_loggingInfo :: Lens' CreateCluster (Maybe LoggingInfo)

Undocumented member.

createCluster_clientAuthentication :: Lens' CreateCluster (Maybe ClientAuthentication)

Includes all client authentication related information.

createCluster_tags :: Lens' CreateCluster (Maybe (HashMap Text Text))

Create tags when creating the cluster.

createCluster_encryptionInfo :: Lens' CreateCluster (Maybe EncryptionInfo)

Includes all encryption-related information.

createCluster_brokerNodeGroupInfo :: Lens' CreateCluster BrokerNodeGroupInfo

Information about the broker nodes in the cluster.

createCluster_kafkaVersion :: Lens' CreateCluster Text

The version of Apache Kafka.

createCluster_numberOfBrokerNodes :: Lens' CreateCluster Natural

The number of broker nodes in the cluster.

createCluster_clusterName :: Lens' CreateCluster Text

The name of the cluster.

Destructuring the Response

data CreateClusterResponse

See: newCreateClusterResponse smart constructor.

Constructors

CreateClusterResponse'

Fields

state :: Maybe ClusterState

The state of the cluster. The possible states are ACTIVE, CREATING, DELETING, FAILED, HEALING, MAINTENANCE, REBOOTING_BROKER, and UPDATING.

clusterArn :: Maybe Text

The Amazon Resource Name (ARN) of the cluster.

clusterName :: Maybe Text

The name of the MSK cluster.

httpStatus :: Int

The response's http status code.

newCreateClusterResponse

Arguments

:: Int
$sel:httpStatus:CreateClusterResponse'

-> CreateClusterResponse

Create a value of CreateClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:CreateClusterResponse', createClusterResponse_state - The state of the cluster. The possible states are ACTIVE, CREATING, DELETING, FAILED, HEALING, MAINTENANCE, REBOOTING_BROKER, and UPDATING.

$sel:clusterArn:CreateClusterResponse', createClusterResponse_clusterArn - The Amazon Resource Name (ARN) of the cluster.

$sel:clusterName:CreateCluster', createClusterResponse_clusterName - The name of the MSK cluster.

$sel:httpStatus:CreateClusterResponse', createClusterResponse_httpStatus - The response's http status code.

Response Lenses

createClusterResponse_state :: Lens' CreateClusterResponse (Maybe ClusterState)

The state of the cluster. The possible states are ACTIVE, CREATING, DELETING, FAILED, HEALING, MAINTENANCE, REBOOTING_BROKER, and UPDATING.

createClusterResponse_clusterArn :: Lens' CreateClusterResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

createClusterResponse_clusterName :: Lens' CreateClusterResponse (Maybe Text)

The name of the MSK cluster.

createClusterResponse_httpStatus :: Lens' CreateClusterResponse Int

The response's http status code.