Network.AWS.Kafka.CreateConfiguration

Description

Creates a new MSK configuration.

Creating a Request

data CreateConfiguration

See: newCreateConfiguration smart constructor.

Constructors

CreateConfiguration'

Fields

kafkaVersions :: Maybe [Text]

The versions of Apache Kafka with which you can use this MSK configuration.

description :: Maybe Text

The description of the configuration.

serverProperties :: Base64

Contents of the server.properties file. When using the API, you must ensure that the contents of the file are base64 encoded. When using the AWS Management Console, the SDK, or the AWS CLI, the contents of server.properties can be in plaintext.

name :: Text

The name of the configuration.

newCreateConfiguration

Arguments

:: ByteString
$sel:serverProperties:CreateConfiguration'

-> Text
$sel:name:CreateConfiguration'

-> CreateConfiguration

Create a value of CreateConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kafkaVersions:CreateConfiguration', createConfiguration_kafkaVersions - The versions of Apache Kafka with which you can use this MSK configuration.

$sel:description:CreateConfiguration', createConfiguration_description - The description of the configuration.

$sel:serverProperties:CreateConfiguration', createConfiguration_serverProperties - Contents of the server.properties file. When using the API, you must ensure that the contents of the file are base64 encoded. When using the AWS Management Console, the SDK, or the AWS CLI, the contents of server.properties can be in plaintext.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:name:CreateConfiguration', createConfiguration_name - The name of the configuration.

Request Lenses

createConfiguration_kafkaVersions :: Lens' CreateConfiguration (Maybe [Text])

The versions of Apache Kafka with which you can use this MSK configuration.

createConfiguration_description :: Lens' CreateConfiguration (Maybe Text)

The description of the configuration.

createConfiguration_serverProperties :: Lens' CreateConfiguration ByteString

Contents of the server.properties file. When using the API, you must ensure that the contents of the file are base64 encoded. When using the AWS Management Console, the SDK, or the AWS CLI, the contents of server.properties can be in plaintext.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

createConfiguration_name :: Lens' CreateConfiguration Text

The name of the configuration.

Destructuring the Response

data CreateConfigurationResponse

See: newCreateConfigurationResponse smart constructor.

Constructors

CreateConfigurationResponse'

Fields

creationTime :: Maybe POSIX

The time when the configuration was created.

state :: Maybe ConfigurationState

The state of the configuration. The possible states are ACTIVE, DELETING, and DELETE_FAILED.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the configuration.

latestRevision :: Maybe ConfigurationRevision

Latest revision of the configuration.

name :: Maybe Text

The name of the configuration.

httpStatus :: Int

The response's http status code.

newCreateConfigurationResponse

Arguments

:: Int
$sel:httpStatus:CreateConfigurationResponse'

-> CreateConfigurationResponse

Create a value of CreateConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:CreateConfigurationResponse', createConfigurationResponse_creationTime - The time when the configuration was created.

$sel:state:CreateConfigurationResponse', createConfigurationResponse_state - The state of the configuration. The possible states are ACTIVE, DELETING, and DELETE_FAILED.

$sel:arn:CreateConfigurationResponse', createConfigurationResponse_arn - The Amazon Resource Name (ARN) of the configuration.

$sel:latestRevision:CreateConfigurationResponse', createConfigurationResponse_latestRevision - Latest revision of the configuration.

$sel:name:CreateConfiguration', createConfigurationResponse_name - The name of the configuration.

$sel:httpStatus:CreateConfigurationResponse', createConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

createConfigurationResponse_creationTime :: Lens' CreateConfigurationResponse (Maybe UTCTime)

The time when the configuration was created.

createConfigurationResponse_state :: Lens' CreateConfigurationResponse (Maybe ConfigurationState)

The state of the configuration. The possible states are ACTIVE, DELETING, and DELETE_FAILED.

createConfigurationResponse_arn :: Lens' CreateConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the configuration.

createConfigurationResponse_latestRevision :: Lens' CreateConfigurationResponse (Maybe ConfigurationRevision)

Latest revision of the configuration.

createConfigurationResponse_name :: Lens' CreateConfigurationResponse (Maybe Text)

The name of the configuration.

createConfigurationResponse_httpStatus :: Lens' CreateConfigurationResponse Int

The response's http status code.