Network.AWS.Kafka.DescribeConfigurationRevision

Description

Returns a description of this revision of the configuration.

Creating a Request

data DescribeConfigurationRevision

See: newDescribeConfigurationRevision smart constructor.

Constructors

DescribeConfigurationRevision'

Fields

revision :: Integer

A string that uniquely identifies a revision of an MSK configuration.

arn :: Text

The Amazon Resource Name (ARN) that uniquely identifies an MSK configuration and all of its revisions.

newDescribeConfigurationRevision

Arguments

:: Integer
$sel:revision:DescribeConfigurationRevision'

-> Text
$sel:arn:DescribeConfigurationRevision'

-> DescribeConfigurationRevision

Create a value of DescribeConfigurationRevision with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:revision:DescribeConfigurationRevision', describeConfigurationRevision_revision - A string that uniquely identifies a revision of an MSK configuration.

$sel:arn:DescribeConfigurationRevision', describeConfigurationRevision_arn - The Amazon Resource Name (ARN) that uniquely identifies an MSK configuration and all of its revisions.

Request Lenses

describeConfigurationRevision_revision :: Lens' DescribeConfigurationRevision Integer

A string that uniquely identifies a revision of an MSK configuration.

describeConfigurationRevision_arn :: Lens' DescribeConfigurationRevision Text

The Amazon Resource Name (ARN) that uniquely identifies an MSK configuration and all of its revisions.

Destructuring the Response

data DescribeConfigurationRevisionResponse

See: newDescribeConfigurationRevisionResponse smart constructor.

Constructors

DescribeConfigurationRevisionResponse'

Fields

creationTime :: Maybe POSIX

The time when the configuration was created.

serverProperties :: Maybe Base64

Contents of the server.properties file. When using the API, you must ensure that the contents of the file are base64 encoded. When using the AWS Management Console, the SDK, or the AWS CLI, the contents of server.properties can be in plaintext.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the configuration.

revision :: Maybe Integer

The revision number.

description :: Maybe Text

The description of the configuration.

httpStatus :: Int

The response's http status code.

newDescribeConfigurationRevisionResponse

Arguments

:: Int
$sel:httpStatus:DescribeConfigurationRevisionResponse'

-> DescribeConfigurationRevisionResponse

Create a value of DescribeConfigurationRevisionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeConfigurationRevisionResponse', describeConfigurationRevisionResponse_creationTime - The time when the configuration was created.

$sel:serverProperties:DescribeConfigurationRevisionResponse', describeConfigurationRevisionResponse_serverProperties - Contents of the server.properties file. When using the API, you must ensure that the contents of the file are base64 encoded. When using the AWS Management Console, the SDK, or the AWS CLI, the contents of server.properties can be in plaintext.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:arn:DescribeConfigurationRevision', describeConfigurationRevisionResponse_arn - The Amazon Resource Name (ARN) of the configuration.

$sel:revision:DescribeConfigurationRevision', describeConfigurationRevisionResponse_revision - The revision number.

$sel:description:DescribeConfigurationRevisionResponse', describeConfigurationRevisionResponse_description - The description of the configuration.

$sel:httpStatus:DescribeConfigurationRevisionResponse', describeConfigurationRevisionResponse_httpStatus - The response's http status code.

Response Lenses

describeConfigurationRevisionResponse_creationTime :: Lens' DescribeConfigurationRevisionResponse (Maybe UTCTime)

The time when the configuration was created.

describeConfigurationRevisionResponse_serverProperties :: Lens' DescribeConfigurationRevisionResponse (Maybe ByteString)

Contents of the server.properties file. When using the API, you must ensure that the contents of the file are base64 encoded. When using the AWS Management Console, the SDK, or the AWS CLI, the contents of server.properties can be in plaintext.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

describeConfigurationRevisionResponse_arn :: Lens' DescribeConfigurationRevisionResponse (Maybe Text)

The Amazon Resource Name (ARN) of the configuration.

describeConfigurationRevisionResponse_revision :: Lens' DescribeConfigurationRevisionResponse (Maybe Integer)

The revision number.

describeConfigurationRevisionResponse_description :: Lens' DescribeConfigurationRevisionResponse (Maybe Text)

The description of the configuration.

describeConfigurationRevisionResponse_httpStatus :: Lens' DescribeConfigurationRevisionResponse Int

The response's http status code.