Network.AWS.Kafka.Lens

Description

Operations

CreateConfiguration

createConfiguration_kafkaVersions :: Lens' CreateConfiguration (Maybe [Text])

The versions of Apache Kafka with which you can use this MSK configuration.

createConfiguration_description :: Lens' CreateConfiguration (Maybe Text)

The description of the configuration.

createConfiguration_serverProperties :: Lens' CreateConfiguration ByteString

Contents of the server.properties file. When using the API, you must ensure that the contents of the file are base64 encoded. When using the AWS Management Console, the SDK, or the AWS CLI, the contents of server.properties can be in plaintext.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

createConfiguration_name :: Lens' CreateConfiguration Text

The name of the configuration.

createConfigurationResponse_creationTime :: Lens' CreateConfigurationResponse (Maybe UTCTime)

The time when the configuration was created.

createConfigurationResponse_state :: Lens' CreateConfigurationResponse (Maybe ConfigurationState)

The state of the configuration. The possible states are ACTIVE, DELETING, and DELETE_FAILED.

createConfigurationResponse_arn :: Lens' CreateConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the configuration.

createConfigurationResponse_latestRevision :: Lens' CreateConfigurationResponse (Maybe ConfigurationRevision)

Latest revision of the configuration.

createConfigurationResponse_name :: Lens' CreateConfigurationResponse (Maybe Text)

The name of the configuration.

createConfigurationResponse_httpStatus :: Lens' CreateConfigurationResponse Int

The response's http status code.

DescribeCluster

describeCluster_clusterArn :: Lens' DescribeCluster Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

describeClusterResponse_clusterInfo :: Lens' DescribeClusterResponse (Maybe ClusterInfo)

The cluster information.

describeClusterResponse_httpStatus :: Lens' DescribeClusterResponse Int

The response's http status code.

RebootBroker

rebootBroker_clusterArn :: Lens' RebootBroker Text

The Amazon Resource Name (ARN) of the cluster to be updated.

rebootBroker_brokerIds :: Lens' RebootBroker [Text]

The list of broker IDs to be rebooted. The reboot-broker operation supports rebooting one broker at a time.

rebootBrokerResponse_clusterArn :: Lens' RebootBrokerResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

rebootBrokerResponse_clusterOperationArn :: Lens' RebootBrokerResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster operation.

rebootBrokerResponse_httpStatus :: Lens' RebootBrokerResponse Int

The response's http status code.

ListConfigurationRevisions

listConfigurationRevisions_nextToken :: Lens' ListConfigurationRevisions (Maybe Text)

The paginated results marker. When the result of the operation is truncated, the call returns NextToken in the response. To get the next batch, provide this token in your next request.

listConfigurationRevisions_maxResults :: Lens' ListConfigurationRevisions (Maybe Natural)

The maximum number of results to return in the response. If there are more results, the response includes a NextToken parameter.

listConfigurationRevisions_arn :: Lens' ListConfigurationRevisions Text

The Amazon Resource Name (ARN) that uniquely identifies an MSK configuration and all of its revisions.

listConfigurationRevisionsResponse_nextToken :: Lens' ListConfigurationRevisionsResponse (Maybe Text)

Paginated results marker.

listConfigurationRevisionsResponse_revisions :: Lens' ListConfigurationRevisionsResponse (Maybe [ConfigurationRevision])

List of ConfigurationRevision objects.

listConfigurationRevisionsResponse_httpStatus :: Lens' ListConfigurationRevisionsResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) that uniquely identifies the resource that's associated with the tags.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The key-value pair for the resource tag.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

ListKafkaVersions

listKafkaVersions_nextToken :: Lens' ListKafkaVersions (Maybe Text)

The paginated results marker. When the result of the operation is truncated, the call returns NextToken in the response. To get the next batch, provide this token in your next request.

listKafkaVersions_maxResults :: Lens' ListKafkaVersions (Maybe Natural)

The maximum number of results to return in the response. If there are more results, the response includes a NextToken parameter.

listKafkaVersionsResponse_kafkaVersions :: Lens' ListKafkaVersionsResponse (Maybe [KafkaVersion])

Undocumented member.

listKafkaVersionsResponse_nextToken :: Lens' ListKafkaVersionsResponse (Maybe Text)

Undocumented member.

listKafkaVersionsResponse_httpStatus :: Lens' ListKafkaVersionsResponse Int

The response's http status code.

UpdateMonitoring

updateMonitoring_enhancedMonitoring :: Lens' UpdateMonitoring (Maybe EnhancedMonitoring)

Specifies which Apache Kafka metrics Amazon MSK gathers and sends to Amazon CloudWatch for this cluster.

updateMonitoring_openMonitoring :: Lens' UpdateMonitoring (Maybe OpenMonitoringInfo)

The settings for open monitoring.

updateMonitoring_loggingInfo :: Lens' UpdateMonitoring (Maybe LoggingInfo)

Undocumented member.

updateMonitoring_clusterArn :: Lens' UpdateMonitoring Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

updateMonitoring_currentVersion :: Lens' UpdateMonitoring Text

The version of the MSK cluster to update. Cluster versions aren't simple numbers. You can describe an MSK cluster to find its version. When this update operation is successful, it generates a new cluster version.

updateMonitoringResponse_clusterArn :: Lens' UpdateMonitoringResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

updateMonitoringResponse_clusterOperationArn :: Lens' UpdateMonitoringResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster operation.

updateMonitoringResponse_httpStatus :: Lens' UpdateMonitoringResponse Int

The response's http status code.

BatchAssociateScramSecret

batchAssociateScramSecret_clusterArn :: Lens' BatchAssociateScramSecret Text

The Amazon Resource Name (ARN) of the cluster to be updated.

batchAssociateScramSecret_secretArnList :: Lens' BatchAssociateScramSecret [Text]

List of AWS Secrets Manager secret ARNs.

batchAssociateScramSecretResponse_clusterArn :: Lens' BatchAssociateScramSecretResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

batchAssociateScramSecretResponse_unprocessedScramSecrets :: Lens' BatchAssociateScramSecretResponse (Maybe [UnprocessedScramSecret])

List of errors when associating secrets to cluster.

batchAssociateScramSecretResponse_httpStatus :: Lens' BatchAssociateScramSecretResponse Int

The response's http status code.

UpdateBrokerStorage

updateBrokerStorage_clusterArn :: Lens' UpdateBrokerStorage Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

updateBrokerStorage_targetBrokerEBSVolumeInfo :: Lens' UpdateBrokerStorage [BrokerEBSVolumeInfo]

Describes the target volume size and the ID of the broker to apply the update to.

updateBrokerStorage_currentVersion :: Lens' UpdateBrokerStorage Text

The version of cluster to update from. A successful operation will then generate a new version.

updateBrokerStorageResponse_clusterArn :: Lens' UpdateBrokerStorageResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

updateBrokerStorageResponse_clusterOperationArn :: Lens' UpdateBrokerStorageResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster operation.

updateBrokerStorageResponse_httpStatus :: Lens' UpdateBrokerStorageResponse Int

The response's http status code.

DeleteCluster

deleteCluster_currentVersion :: Lens' DeleteCluster (Maybe Text)

The current version of the MSK cluster.

deleteCluster_clusterArn :: Lens' DeleteCluster Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

deleteClusterResponse_state :: Lens' DeleteClusterResponse (Maybe ClusterState)

The state of the cluster. The possible states are ACTIVE, CREATING, DELETING, FAILED, HEALING, MAINTENANCE, REBOOTING_BROKER, and UPDATING.

deleteClusterResponse_clusterArn :: Lens' DeleteClusterResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

deleteClusterResponse_httpStatus :: Lens' DeleteClusterResponse Int

The response's http status code.

UpdateClusterConfiguration

updateClusterConfiguration_clusterArn :: Lens' UpdateClusterConfiguration Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

updateClusterConfiguration_currentVersion :: Lens' UpdateClusterConfiguration Text

The version of the cluster that needs to be updated.

updateClusterConfiguration_configurationInfo :: Lens' UpdateClusterConfiguration ConfigurationInfo

Represents the configuration that you want MSK to use for the brokers in a cluster.

updateClusterConfigurationResponse_clusterArn :: Lens' UpdateClusterConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

updateClusterConfigurationResponse_clusterOperationArn :: Lens' UpdateClusterConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster operation.

updateClusterConfigurationResponse_httpStatus :: Lens' UpdateClusterConfigurationResponse Int

The response's http status code.

CreateCluster

createCluster_enhancedMonitoring :: Lens' CreateCluster (Maybe EnhancedMonitoring)

Specifies the level of monitoring for the MSK cluster. The possible values are DEFAULT, PER_BROKER, PER_TOPIC_PER_BROKER, and PER_TOPIC_PER_PARTITION.

createCluster_openMonitoring :: Lens' CreateCluster (Maybe OpenMonitoringInfo)

The settings for open monitoring.

createCluster_configurationInfo :: Lens' CreateCluster (Maybe ConfigurationInfo)

Represents the configuration that you want MSK to use for the brokers in a cluster.

createCluster_loggingInfo :: Lens' CreateCluster (Maybe LoggingInfo)

Undocumented member.

createCluster_clientAuthentication :: Lens' CreateCluster (Maybe ClientAuthentication)

Includes all client authentication related information.

createCluster_tags :: Lens' CreateCluster (Maybe (HashMap Text Text))

Create tags when creating the cluster.

createCluster_encryptionInfo :: Lens' CreateCluster (Maybe EncryptionInfo)

Includes all encryption-related information.

createCluster_brokerNodeGroupInfo :: Lens' CreateCluster BrokerNodeGroupInfo

Information about the broker nodes in the cluster.

createCluster_kafkaVersion :: Lens' CreateCluster Text

The version of Apache Kafka.

createCluster_numberOfBrokerNodes :: Lens' CreateCluster Natural

The number of broker nodes in the cluster.

createCluster_clusterName :: Lens' CreateCluster Text

The name of the cluster.

createClusterResponse_state :: Lens' CreateClusterResponse (Maybe ClusterState)

The state of the cluster. The possible states are ACTIVE, CREATING, DELETING, FAILED, HEALING, MAINTENANCE, REBOOTING_BROKER, and UPDATING.

createClusterResponse_clusterArn :: Lens' CreateClusterResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

createClusterResponse_clusterName :: Lens' CreateClusterResponse (Maybe Text)

The name of the MSK cluster.

createClusterResponse_httpStatus :: Lens' CreateClusterResponse Int

The response's http status code.

UpdateBrokerCount

updateBrokerCount_clusterArn :: Lens' UpdateBrokerCount Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

updateBrokerCount_currentVersion :: Lens' UpdateBrokerCount Text

The version of cluster to update from. A successful operation will then generate a new version.

updateBrokerCount_targetNumberOfBrokerNodes :: Lens' UpdateBrokerCount Natural

The number of broker nodes that you want the cluster to have after this operation completes successfully.

updateBrokerCountResponse_clusterArn :: Lens' UpdateBrokerCountResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

updateBrokerCountResponse_clusterOperationArn :: Lens' UpdateBrokerCountResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster operation.

updateBrokerCountResponse_httpStatus :: Lens' UpdateBrokerCountResponse Int

The response's http status code.

ListConfigurations

listConfigurations_nextToken :: Lens' ListConfigurations (Maybe Text)

The paginated results marker. When the result of the operation is truncated, the call returns NextToken in the response. To get the next batch, provide this token in your next request.

listConfigurations_maxResults :: Lens' ListConfigurations (Maybe Natural)

The maximum number of results to return in the response. If there are more results, the response includes a NextToken parameter.

listConfigurationsResponse_configurations :: Lens' ListConfigurationsResponse (Maybe [Configuration])

An array of MSK configurations.

listConfigurationsResponse_nextToken :: Lens' ListConfigurationsResponse (Maybe Text)

The paginated results marker. When the result of a ListConfigurations operation is truncated, the call returns NextToken in the response. To get another batch of configurations, provide this token in your next request.

listConfigurationsResponse_httpStatus :: Lens' ListConfigurationsResponse Int

The response's http status code.

GetBootstrapBrokers

getBootstrapBrokers_clusterArn :: Lens' GetBootstrapBrokers Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

getBootstrapBrokersResponse_bootstrapBrokerString :: Lens' GetBootstrapBrokersResponse (Maybe Text)

A string containing one or more hostname:port pairs.

getBootstrapBrokersResponse_bootstrapBrokerStringSaslScram :: Lens' GetBootstrapBrokersResponse (Maybe Text)

A string containing one or more DNS names (or IP) and Sasl Scram port pairs.

getBootstrapBrokersResponse_bootstrapBrokerStringTls :: Lens' GetBootstrapBrokersResponse (Maybe Text)

A string containing one or more DNS names (or IP) and TLS port pairs.

getBootstrapBrokersResponse_bootstrapBrokerStringSaslIam :: Lens' GetBootstrapBrokersResponse (Maybe Text)

A string that contains one or more DNS names (or IP addresses) and SASL IAM port pairs.

getBootstrapBrokersResponse_httpStatus :: Lens' GetBootstrapBrokersResponse Int

The response's http status code.

UpdateClusterKafkaVersion

updateClusterKafkaVersion_configurationInfo :: Lens' UpdateClusterKafkaVersion (Maybe ConfigurationInfo)

The custom configuration that should be applied on the new version of cluster.

updateClusterKafkaVersion_clusterArn :: Lens' UpdateClusterKafkaVersion Text

The Amazon Resource Name (ARN) of the cluster to be updated.

updateClusterKafkaVersion_targetKafkaVersion :: Lens' UpdateClusterKafkaVersion Text

Target Kafka version.

updateClusterKafkaVersion_currentVersion :: Lens' UpdateClusterKafkaVersion Text

Current cluster version.

updateClusterKafkaVersionResponse_clusterArn :: Lens' UpdateClusterKafkaVersionResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

updateClusterKafkaVersionResponse_clusterOperationArn :: Lens' UpdateClusterKafkaVersionResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster operation.

updateClusterKafkaVersionResponse_httpStatus :: Lens' UpdateClusterKafkaVersionResponse Int

The response's http status code.

UpdateSecurity

updateSecurity_clientAuthentication :: Lens' UpdateSecurity (Maybe ClientAuthentication)

Includes all client authentication related information.

updateSecurity_encryptionInfo :: Lens' UpdateSecurity (Maybe EncryptionInfo)

Includes all encryption-related information.

updateSecurity_clusterArn :: Lens' UpdateSecurity Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

updateSecurity_currentVersion :: Lens' UpdateSecurity Text

The version of the MSK cluster to update. Cluster versions aren't simple numbers. You can describe an MSK cluster to find its version. When this update operation is successful, it generates a new cluster version.

updateSecurityResponse_clusterArn :: Lens' UpdateSecurityResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

updateSecurityResponse_clusterOperationArn :: Lens' UpdateSecurityResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster operation.

updateSecurityResponse_httpStatus :: Lens' UpdateSecurityResponse Int

The response's http status code.

GetCompatibleKafkaVersions

getCompatibleKafkaVersions_clusterArn :: Lens' GetCompatibleKafkaVersions (Maybe Text)

The Amazon Resource Name (ARN) of the cluster check.

getCompatibleKafkaVersionsResponse_compatibleKafkaVersions :: Lens' GetCompatibleKafkaVersionsResponse (Maybe [CompatibleKafkaVersion])

A list of CompatibleKafkaVersion objects.

getCompatibleKafkaVersionsResponse_httpStatus :: Lens' GetCompatibleKafkaVersionsResponse Int

The response's http status code.

DescribeClusterOperation

describeClusterOperation_clusterOperationArn :: Lens' DescribeClusterOperation Text

The Amazon Resource Name (ARN) that uniquely identifies the MSK cluster operation.

describeClusterOperationResponse_clusterOperationInfo :: Lens' DescribeClusterOperationResponse (Maybe ClusterOperationInfo)

Cluster operation information

describeClusterOperationResponse_httpStatus :: Lens' DescribeClusterOperationResponse Int

The response's http status code.

UpdateBrokerType

updateBrokerType_clusterArn :: Lens' UpdateBrokerType Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

updateBrokerType_currentVersion :: Lens' UpdateBrokerType Text

The cluster version that you want to change. After this operation completes successfully, the cluster will have a new version.

updateBrokerType_targetInstanceType :: Lens' UpdateBrokerType Text

The Amazon MSK broker type that you want all of the brokers in this cluster to be.

updateBrokerTypeResponse_clusterArn :: Lens' UpdateBrokerTypeResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

updateBrokerTypeResponse_clusterOperationArn :: Lens' UpdateBrokerTypeResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster operation.

updateBrokerTypeResponse_httpStatus :: Lens' UpdateBrokerTypeResponse Int

The response's http status code.

DescribeConfiguration

describeConfiguration_arn :: Lens' DescribeConfiguration Text

The Amazon Resource Name (ARN) that uniquely identifies an MSK configuration and all of its revisions.

describeConfigurationResponse_creationTime :: Lens' DescribeConfigurationResponse (Maybe UTCTime)

The time when the configuration was created.

describeConfigurationResponse_state :: Lens' DescribeConfigurationResponse (Maybe ConfigurationState)

The state of the configuration. The possible states are ACTIVE, DELETING, and DELETE_FAILED.

describeConfigurationResponse_kafkaVersions :: Lens' DescribeConfigurationResponse (Maybe [Text])

The versions of Apache Kafka with which you can use this MSK configuration.

describeConfigurationResponse_arn :: Lens' DescribeConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the configuration.

describeConfigurationResponse_latestRevision :: Lens' DescribeConfigurationResponse (Maybe ConfigurationRevision)

Latest revision of the configuration.

describeConfigurationResponse_name :: Lens' DescribeConfigurationResponse (Maybe Text)

The name of the configuration.

describeConfigurationResponse_description :: Lens' DescribeConfigurationResponse (Maybe Text)

The description of the configuration.

describeConfigurationResponse_httpStatus :: Lens' DescribeConfigurationResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) that uniquely identifies the resource that's associated with the tags.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The key-value pair for the resource tag.

ListClusters

listClusters_clusterNameFilter :: Lens' ListClusters (Maybe Text)

Specify a prefix of the name of the clusters that you want to list. The service lists all the clusters whose names start with this prefix.

listClusters_nextToken :: Lens' ListClusters (Maybe Text)

The paginated results marker. When the result of the operation is truncated, the call returns NextToken in the response. To get the next batch, provide this token in your next request.

listClusters_maxResults :: Lens' ListClusters (Maybe Natural)

The maximum number of results to return in the response. If there are more results, the response includes a NextToken parameter.

listClustersResponse_nextToken :: Lens' ListClustersResponse (Maybe Text)

The paginated results marker. When the result of a ListClusters operation is truncated, the call returns NextToken in the response. To get another batch of clusters, provide this token in your next request.

listClustersResponse_clusterInfoList :: Lens' ListClustersResponse (Maybe [ClusterInfo])

Information on each of the MSK clusters in the response.

listClustersResponse_httpStatus :: Lens' ListClustersResponse Int

The response's http status code.

UntagResource

untagResource_tagKeys :: Lens' UntagResource [Text]

Tag keys must be unique for a given cluster. In addition, the following restrictions apply:

Each tag key must be unique. If you add a tag with a key that's already in use, your new tag overwrites the existing key-value pair.

You can't start a tag key with aws: because this prefix is reserved for use by AWS. AWS creates tags that begin with this prefix on your behalf, but you can't edit or delete them.

Tag keys must be between 1 and 128 Unicode characters in length.

Tag keys must consist of the following characters: Unicode letters, digits, white space, and the following special characters: _ . / = + - @.

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) that uniquely identifies the resource that's associated with the tags.

ListClusterOperations

listClusterOperations_nextToken :: Lens' ListClusterOperations (Maybe Text)

The paginated results marker. When the result of the operation is truncated, the call returns NextToken in the response. To get the next batch, provide this token in your next request.

listClusterOperations_maxResults :: Lens' ListClusterOperations (Maybe Natural)

The maximum number of results to return in the response. If there are more results, the response includes a NextToken parameter.

listClusterOperations_clusterArn :: Lens' ListClusterOperations Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

listClusterOperationsResponse_clusterOperationInfoList :: Lens' ListClusterOperationsResponse (Maybe [ClusterOperationInfo])

An array of cluster operation information objects.

listClusterOperationsResponse_nextToken :: Lens' ListClusterOperationsResponse (Maybe Text)

If the response of ListClusterOperations is truncated, it returns a NextToken in the response. This Nexttoken should be sent in the subsequent request to ListClusterOperations.

listClusterOperationsResponse_httpStatus :: Lens' ListClusterOperationsResponse Int

The response's http status code.

BatchDisassociateScramSecret

batchDisassociateScramSecret_clusterArn :: Lens' BatchDisassociateScramSecret Text

The Amazon Resource Name (ARN) of the cluster to be updated.

batchDisassociateScramSecret_secretArnList :: Lens' BatchDisassociateScramSecret [Text]

List of AWS Secrets Manager secret ARNs.

batchDisassociateScramSecretResponse_clusterArn :: Lens' BatchDisassociateScramSecretResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

batchDisassociateScramSecretResponse_unprocessedScramSecrets :: Lens' BatchDisassociateScramSecretResponse (Maybe [UnprocessedScramSecret])

List of errors when disassociating secrets to cluster.

batchDisassociateScramSecretResponse_httpStatus :: Lens' BatchDisassociateScramSecretResponse Int

The response's http status code.

DescribeConfigurationRevision

describeConfigurationRevision_revision :: Lens' DescribeConfigurationRevision Integer

A string that uniquely identifies a revision of an MSK configuration.

describeConfigurationRevision_arn :: Lens' DescribeConfigurationRevision Text

The Amazon Resource Name (ARN) that uniquely identifies an MSK configuration and all of its revisions.

describeConfigurationRevisionResponse_creationTime :: Lens' DescribeConfigurationRevisionResponse (Maybe UTCTime)

The time when the configuration was created.

describeConfigurationRevisionResponse_serverProperties :: Lens' DescribeConfigurationRevisionResponse (Maybe ByteString)

Contents of the server.properties file. When using the API, you must ensure that the contents of the file are base64 encoded. When using the AWS Management Console, the SDK, or the AWS CLI, the contents of server.properties can be in plaintext.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

describeConfigurationRevisionResponse_arn :: Lens' DescribeConfigurationRevisionResponse (Maybe Text)

The Amazon Resource Name (ARN) of the configuration.

describeConfigurationRevisionResponse_revision :: Lens' DescribeConfigurationRevisionResponse (Maybe Integer)

The revision number.

describeConfigurationRevisionResponse_description :: Lens' DescribeConfigurationRevisionResponse (Maybe Text)

The description of the configuration.

describeConfigurationRevisionResponse_httpStatus :: Lens' DescribeConfigurationRevisionResponse Int

The response's http status code.

DeleteConfiguration

deleteConfiguration_arn :: Lens' DeleteConfiguration Text

The Amazon Resource Name (ARN) that uniquely identifies an MSK configuration.

deleteConfigurationResponse_state :: Lens' DeleteConfigurationResponse (Maybe ConfigurationState)

The state of the configuration. The possible states are ACTIVE, DELETING, and DELETE_FAILED.

deleteConfigurationResponse_arn :: Lens' DeleteConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) that uniquely identifies an MSK configuration.

deleteConfigurationResponse_httpStatus :: Lens' DeleteConfigurationResponse Int

The response's http status code.

UpdateConfiguration

updateConfiguration_description :: Lens' UpdateConfiguration (Maybe Text)

The description of the configuration revision.

updateConfiguration_arn :: Lens' UpdateConfiguration Text

The Amazon Resource Name (ARN) of the configuration.

updateConfiguration_serverProperties :: Lens' UpdateConfiguration ByteString

Contents of the server.properties file. When using the API, you must ensure that the contents of the file are base64 encoded. When using the AWS Management Console, the SDK, or the AWS CLI, the contents of server.properties can be in plaintext.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

updateConfigurationResponse_arn :: Lens' UpdateConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the configuration.

updateConfigurationResponse_latestRevision :: Lens' UpdateConfigurationResponse (Maybe ConfigurationRevision)

Latest revision of the configuration.

updateConfigurationResponse_httpStatus :: Lens' UpdateConfigurationResponse Int

The response's http status code.

ListNodes

listNodes_nextToken :: Lens' ListNodes (Maybe Text)

The paginated results marker. When the result of the operation is truncated, the call returns NextToken in the response. To get the next batch, provide this token in your next request.

listNodes_maxResults :: Lens' ListNodes (Maybe Natural)

The maximum number of results to return in the response. If there are more results, the response includes a NextToken parameter.

listNodes_clusterArn :: Lens' ListNodes Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

listNodesResponse_nodeInfoList :: Lens' ListNodesResponse (Maybe [NodeInfo])

List containing a NodeInfo object.

listNodesResponse_nextToken :: Lens' ListNodesResponse (Maybe Text)

The paginated results marker. When the result of a ListNodes operation is truncated, the call returns NextToken in the response. To get another batch of nodes, provide this token in your next request.

listNodesResponse_httpStatus :: Lens' ListNodesResponse Int

The response's http status code.

ListScramSecrets

listScramSecrets_nextToken :: Lens' ListScramSecrets (Maybe Text)

The nextToken of the query.

listScramSecrets_maxResults :: Lens' ListScramSecrets (Maybe Natural)

The maxResults of the query.

listScramSecrets_clusterArn :: Lens' ListScramSecrets Text

The arn of the cluster.

listScramSecretsResponse_nextToken :: Lens' ListScramSecretsResponse (Maybe Text)

Paginated results marker.

listScramSecretsResponse_secretArnList :: Lens' ListScramSecretsResponse (Maybe [Text])

The list of scram secrets associated with the cluster.

listScramSecretsResponse_httpStatus :: Lens' ListScramSecretsResponse Int

The response's http status code.

Types

BrokerEBSVolumeInfo

brokerEBSVolumeInfo_volumeSizeGB :: Lens' BrokerEBSVolumeInfo Int

Size of the EBS volume to update.

brokerEBSVolumeInfo_kafkaBrokerNodeId :: Lens' BrokerEBSVolumeInfo Text

The ID of the broker to update.

BrokerLogs

brokerLogs_cloudWatchLogs :: Lens' BrokerLogs (Maybe CloudWatchLogs)

Undocumented member.

brokerLogs_firehose :: Lens' BrokerLogs (Maybe Firehose)

Undocumented member.

brokerLogs_s3 :: Lens' BrokerLogs (Maybe S3)

Undocumented member.

BrokerNodeGroupInfo

brokerNodeGroupInfo_storageInfo :: Lens' BrokerNodeGroupInfo (Maybe StorageInfo)

Contains information about storage volumes attached to MSK broker nodes.

brokerNodeGroupInfo_brokerAZDistribution :: Lens' BrokerNodeGroupInfo (Maybe BrokerAZDistribution)

The distribution of broker nodes across Availability Zones. This is an optional parameter. If you don't specify it, Amazon MSK gives it the value DEFAULT. You can also explicitly set this parameter to the value DEFAULT. No other values are currently allowed.

Amazon MSK distributes the broker nodes evenly across the Availability Zones that correspond to the subnets you provide when you create the cluster.

brokerNodeGroupInfo_securityGroups :: Lens' BrokerNodeGroupInfo (Maybe [Text])

The AWS security groups to associate with the elastic network interfaces in order to specify who can connect to and communicate with the Amazon MSK cluster. If you don't specify a security group, Amazon MSK uses the default security group associated with the VPC.

brokerNodeGroupInfo_clientSubnets :: Lens' BrokerNodeGroupInfo [Text]

The list of subnets to connect to in the client virtual private cloud (VPC). AWS creates elastic network interfaces inside these subnets. Client applications use elastic network interfaces to produce and consume data. Client subnets can't be in Availability Zone us-east-1e.

brokerNodeGroupInfo_instanceType :: Lens' BrokerNodeGroupInfo Text

The type of Amazon EC2 instances to use for Kafka brokers. The following instance types are allowed: kafka.m5.large, kafka.m5.xlarge, kafka.m5.2xlarge, kafka.m5.4xlarge, kafka.m5.12xlarge, and kafka.m5.24xlarge.

BrokerNodeInfo

brokerNodeInfo_currentBrokerSoftwareInfo :: Lens' BrokerNodeInfo (Maybe BrokerSoftwareInfo)

Information about the version of software currently deployed on the Kafka brokers in the cluster.

brokerNodeInfo_clientSubnet :: Lens' BrokerNodeInfo (Maybe Text)

The client subnet to which this broker node belongs.

brokerNodeInfo_attachedENIId :: Lens' BrokerNodeInfo (Maybe Text)

The attached elastic network interface of the broker.

brokerNodeInfo_endpoints :: Lens' BrokerNodeInfo (Maybe [Text])

Endpoints for accessing the broker.

brokerNodeInfo_clientVpcIpAddress :: Lens' BrokerNodeInfo (Maybe Text)

The virtual private cloud (VPC) of the client.

brokerNodeInfo_brokerId :: Lens' BrokerNodeInfo (Maybe Double)

The ID of the broker.

BrokerSoftwareInfo

brokerSoftwareInfo_configurationRevision :: Lens' BrokerSoftwareInfo (Maybe Integer)

The revision of the configuration to use. This field isn't visible in this preview release.

brokerSoftwareInfo_kafkaVersion :: Lens' BrokerSoftwareInfo (Maybe Text)

The version of Apache Kafka.

brokerSoftwareInfo_configurationArn :: Lens' BrokerSoftwareInfo (Maybe Text)

The Amazon Resource Name (ARN) of the configuration used for the cluster. This field isn't visible in this preview release.

ClientAuthentication

clientAuthentication_sasl :: Lens' ClientAuthentication (Maybe Sasl)

Details for ClientAuthentication using SASL.

clientAuthentication_tls :: Lens' ClientAuthentication (Maybe Tls)

Details for ClientAuthentication using TLS.

clientAuthentication_unauthenticated :: Lens' ClientAuthentication (Maybe Unauthenticated)

Contains information about unauthenticated traffic to the cluster.

CloudWatchLogs

cloudWatchLogs_logGroup :: Lens' CloudWatchLogs (Maybe Text)

Undocumented member.

cloudWatchLogs_enabled :: Lens' CloudWatchLogs Bool

Undocumented member.

ClusterInfo

clusterInfo_creationTime :: Lens' ClusterInfo (Maybe UTCTime)

The time when the cluster was created.

clusterInfo_activeOperationArn :: Lens' ClusterInfo (Maybe Text)

Arn of active cluster operation.

clusterInfo_state :: Lens' ClusterInfo (Maybe ClusterState)

The state of the cluster. The possible states are ACTIVE, CREATING, DELETING, FAILED, HEALING, MAINTENANCE, REBOOTING_BROKER, and UPDATING.

clusterInfo_clusterArn :: Lens' ClusterInfo (Maybe Text)

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

clusterInfo_numberOfBrokerNodes :: Lens' ClusterInfo (Maybe Int)

The number of broker nodes in the cluster.

clusterInfo_enhancedMonitoring :: Lens' ClusterInfo (Maybe EnhancedMonitoring)

Specifies which metrics are gathered for the MSK cluster. This property has the following possible values: DEFAULT, PER_BROKER, PER_TOPIC_PER_BROKER, and PER_TOPIC_PER_PARTITION. For a list of the metrics associated with each of these levels of monitoring, see Monitoring.

clusterInfo_brokerNodeGroupInfo :: Lens' ClusterInfo (Maybe BrokerNodeGroupInfo)

Information about the broker nodes.

clusterInfo_openMonitoring :: Lens' ClusterInfo (Maybe OpenMonitoring)

Settings for open monitoring using Prometheus.

clusterInfo_currentBrokerSoftwareInfo :: Lens' ClusterInfo (Maybe BrokerSoftwareInfo)

Information about the version of software currently deployed on the Kafka brokers in the cluster.

clusterInfo_currentVersion :: Lens' ClusterInfo (Maybe Text)

The current version of the MSK cluster.

clusterInfo_stateInfo :: Lens' ClusterInfo (Maybe StateInfo)

Undocumented member.

clusterInfo_loggingInfo :: Lens' ClusterInfo (Maybe LoggingInfo)

Undocumented member.

clusterInfo_zookeeperConnectString :: Lens' ClusterInfo (Maybe Text)

The connection string to use to connect to the Apache ZooKeeper cluster.

clusterInfo_clusterName :: Lens' ClusterInfo (Maybe Text)

The name of the cluster.

clusterInfo_zookeeperConnectStringTls :: Lens' ClusterInfo (Maybe Text)

The connection string to use to connect to zookeeper cluster on Tls port.

clusterInfo_clientAuthentication :: Lens' ClusterInfo (Maybe ClientAuthentication)

Includes all client authentication information.

clusterInfo_tags :: Lens' ClusterInfo (Maybe (HashMap Text Text))

Tags attached to the cluster.

clusterInfo_encryptionInfo :: Lens' ClusterInfo (Maybe EncryptionInfo)

Includes all encryption-related information.

ClusterOperationInfo

clusterOperationInfo_creationTime :: Lens' ClusterOperationInfo (Maybe UTCTime)

The time that the operation was created.

clusterOperationInfo_clusterArn :: Lens' ClusterOperationInfo (Maybe Text)

ARN of the cluster.

clusterOperationInfo_clientRequestId :: Lens' ClusterOperationInfo (Maybe Text)

The ID of the API request that triggered this operation.

clusterOperationInfo_targetClusterInfo :: Lens' ClusterOperationInfo (Maybe MutableClusterInfo)

Information about cluster attributes after a cluster is updated.

clusterOperationInfo_sourceClusterInfo :: Lens' ClusterOperationInfo (Maybe MutableClusterInfo)

Information about cluster attributes before a cluster is updated.

clusterOperationInfo_operationSteps :: Lens' ClusterOperationInfo (Maybe [ClusterOperationStep])

Steps completed during the operation.

clusterOperationInfo_operationState :: Lens' ClusterOperationInfo (Maybe Text)

State of the cluster operation.

clusterOperationInfo_endTime :: Lens' ClusterOperationInfo (Maybe UTCTime)

The time at which the operation finished.

clusterOperationInfo_operationType :: Lens' ClusterOperationInfo (Maybe Text)

Type of the cluster operation.

clusterOperationInfo_operationArn :: Lens' ClusterOperationInfo (Maybe Text)

ARN of the cluster operation.

clusterOperationInfo_errorInfo :: Lens' ClusterOperationInfo (Maybe ErrorInfo)

Describes the error if the operation fails.

ClusterOperationStep

clusterOperationStep_stepName :: Lens' ClusterOperationStep (Maybe Text)

The name of the step.

clusterOperationStep_stepInfo :: Lens' ClusterOperationStep (Maybe ClusterOperationStepInfo)

Information about the step and its status.

ClusterOperationStepInfo

clusterOperationStepInfo_stepStatus :: Lens' ClusterOperationStepInfo (Maybe Text)

The steps current status.

CompatibleKafkaVersion

compatibleKafkaVersion_sourceVersion :: Lens' CompatibleKafkaVersion (Maybe Text)

A Kafka version.

compatibleKafkaVersion_targetVersions :: Lens' CompatibleKafkaVersion (Maybe [Text])

A list of Kafka versions.

Configuration

configuration_description :: Lens' Configuration Text

The description of the configuration.

configuration_latestRevision :: Lens' Configuration ConfigurationRevision

Latest revision of the configuration.

configuration_creationTime :: Lens' Configuration UTCTime

The time when the configuration was created.

configuration_kafkaVersions :: Lens' Configuration [Text]

An array of the versions of Apache Kafka with which you can use this MSK configuration. You can use this configuration for an MSK cluster only if the Apache Kafka version specified for the cluster appears in this array.

configuration_arn :: Lens' Configuration Text

The Amazon Resource Name (ARN) of the configuration.

configuration_name :: Lens' Configuration Text

The name of the configuration.

configuration_state :: Lens' Configuration ConfigurationState

The state of the configuration. The possible states are ACTIVE, DELETING, and DELETE_FAILED.

ConfigurationInfo

configurationInfo_revision :: Lens' ConfigurationInfo Integer

The revision of the configuration to use.

configurationInfo_arn :: Lens' ConfigurationInfo Text

ARN of the configuration to use.

ConfigurationRevision

configurationRevision_description :: Lens' ConfigurationRevision (Maybe Text)

The description of the configuration revision.

configurationRevision_revision :: Lens' ConfigurationRevision Integer

The revision number.

configurationRevision_creationTime :: Lens' ConfigurationRevision UTCTime

The time when the configuration revision was created.

EBSStorageInfo

eBSStorageInfo_volumeSize :: Lens' EBSStorageInfo (Maybe Natural)

The size in GiB of the EBS volume for the data drive on each broker node.

EncryptionAtRest

encryptionAtRest_dataVolumeKMSKeyId :: Lens' EncryptionAtRest Text

The ARN of the AWS KMS key for encrypting data at rest. If you don't specify a KMS key, MSK creates one for you and uses it.

EncryptionInTransit

encryptionInTransit_clientBroker :: Lens' EncryptionInTransit (Maybe ClientBroker)

Indicates the encryption setting for data in transit between clients and brokers. The following are the possible values.

TLS means that client-broker communication is enabled with TLS only.

TLS_PLAINTEXT means that client-broker communication is enabled for both TLS-encrypted, as well as plaintext data.

PLAINTEXT means that client-broker communication is enabled in plaintext only.

The default value is TLS_PLAINTEXT.

encryptionInTransit_inCluster :: Lens' EncryptionInTransit (Maybe Bool)

When set to true, it indicates that data communication among the broker nodes of the cluster is encrypted. When set to false, the communication happens in plaintext.

The default value is true.

EncryptionInfo

encryptionInfo_encryptionAtRest :: Lens' EncryptionInfo (Maybe EncryptionAtRest)

The data-volume encryption details.

encryptionInfo_encryptionInTransit :: Lens' EncryptionInfo (Maybe EncryptionInTransit)

The details for encryption in transit.

ErrorInfo

errorInfo_errorString :: Lens' ErrorInfo (Maybe Text)

An optional field to provide more details about the error.

errorInfo_errorCode :: Lens' ErrorInfo (Maybe Text)

A number describing the error programmatically.

Firehose

firehose_deliveryStream :: Lens' Firehose (Maybe Text)

Undocumented member.

firehose_enabled :: Lens' Firehose Bool

Undocumented member.

Iam

iam_enabled :: Lens' Iam (Maybe Bool)

Indicates whether IAM access control is enabled.

JmxExporter

jmxExporter_enabledInBroker :: Lens' JmxExporter Bool

Indicates whether you want to enable or disable the JMX Exporter.

JmxExporterInfo

jmxExporterInfo_enabledInBroker :: Lens' JmxExporterInfo Bool

Indicates whether you want to enable or disable the JMX Exporter.

KafkaVersion

kafkaVersion_status :: Lens' KafkaVersion (Maybe KafkaVersionStatus)

Undocumented member.

kafkaVersion_version :: Lens' KafkaVersion (Maybe Text)

Undocumented member.

LoggingInfo

loggingInfo_brokerLogs :: Lens' LoggingInfo BrokerLogs

Undocumented member.

MutableClusterInfo

mutableClusterInfo_numberOfBrokerNodes :: Lens' MutableClusterInfo (Maybe Int)

The number of broker nodes in the cluster.

mutableClusterInfo_enhancedMonitoring :: Lens' MutableClusterInfo (Maybe EnhancedMonitoring)

Specifies which Apache Kafka metrics Amazon MSK gathers and sends to Amazon CloudWatch for this cluster.

mutableClusterInfo_openMonitoring :: Lens' MutableClusterInfo (Maybe OpenMonitoring)

The settings for open monitoring.

mutableClusterInfo_configurationInfo :: Lens' MutableClusterInfo (Maybe ConfigurationInfo)

Information about the changes in the configuration of the brokers.

mutableClusterInfo_instanceType :: Lens' MutableClusterInfo (Maybe Text)

Information about the Amazon MSK broker type.

mutableClusterInfo_kafkaVersion :: Lens' MutableClusterInfo (Maybe Text)

The Kafka version.

mutableClusterInfo_loggingInfo :: Lens' MutableClusterInfo (Maybe LoggingInfo)

You can configure your MSK cluster to send broker logs to different destination types. This is a container for the configuration details related to broker logs.

mutableClusterInfo_clientAuthentication :: Lens' MutableClusterInfo (Maybe ClientAuthentication)

Includes all client authentication information.

mutableClusterInfo_brokerEBSVolumeInfo :: Lens' MutableClusterInfo (Maybe [BrokerEBSVolumeInfo])

Specifies the size of the EBS volume and the ID of the associated broker.

mutableClusterInfo_encryptionInfo :: Lens' MutableClusterInfo (Maybe EncryptionInfo)

Includes all encryption-related information.

NodeExporter

nodeExporter_enabledInBroker :: Lens' NodeExporter Bool

Indicates whether you want to enable or disable the Node Exporter.

NodeExporterInfo

nodeExporterInfo_enabledInBroker :: Lens' NodeExporterInfo Bool

Indicates whether you want to enable or disable the Node Exporter.

NodeInfo

nodeInfo_addedToClusterTime :: Lens' NodeInfo (Maybe Text)

The start time.

nodeInfo_nodeARN :: Lens' NodeInfo (Maybe Text)

The Amazon Resource Name (ARN) of the node.

nodeInfo_zookeeperNodeInfo :: Lens' NodeInfo (Maybe ZookeeperNodeInfo)

The ZookeeperNodeInfo.

nodeInfo_instanceType :: Lens' NodeInfo (Maybe Text)

The instance type.

nodeInfo_brokerNodeInfo :: Lens' NodeInfo (Maybe BrokerNodeInfo)

The broker node info.

nodeInfo_nodeType :: Lens' NodeInfo (Maybe NodeType)

The node type.

OpenMonitoring

openMonitoring_prometheus :: Lens' OpenMonitoring Prometheus

Prometheus settings.

OpenMonitoringInfo

openMonitoringInfo_prometheus :: Lens' OpenMonitoringInfo PrometheusInfo

Prometheus settings.

Prometheus

prometheus_jmxExporter :: Lens' Prometheus (Maybe JmxExporter)

Indicates whether you want to enable or disable the JMX Exporter.

prometheus_nodeExporter :: Lens' Prometheus (Maybe NodeExporter)

Indicates whether you want to enable or disable the Node Exporter.

PrometheusInfo

prometheusInfo_jmxExporter :: Lens' PrometheusInfo (Maybe JmxExporterInfo)

Indicates whether you want to enable or disable the JMX Exporter.

prometheusInfo_nodeExporter :: Lens' PrometheusInfo (Maybe NodeExporterInfo)

Indicates whether you want to enable or disable the Node Exporter.

S3

s3_prefix :: Lens' S3 (Maybe Text)

Undocumented member.

s3_bucket :: Lens' S3 (Maybe Text)

Undocumented member.

s3_enabled :: Lens' S3 Bool

Undocumented member.

Sasl

sasl_iam :: Lens' Sasl (Maybe Iam)

Indicates whether IAM access control is enabled.

sasl_scram :: Lens' Sasl (Maybe Scram)

Details for SASL/SCRAM client authentication.

Scram

scram_enabled :: Lens' Scram (Maybe Bool)

SASL/SCRAM authentication is enabled or not.

StateInfo

stateInfo_code :: Lens' StateInfo (Maybe Text)

Undocumented member.

stateInfo_message :: Lens' StateInfo (Maybe Text)

Undocumented member.

StorageInfo

storageInfo_ebsStorageInfo :: Lens' StorageInfo (Maybe EBSStorageInfo)

EBS volume information.

Tls

tls_enabled :: Lens' Tls (Maybe Bool)

Specifies whether you want to enable or disable TLS authentication.

tls_certificateAuthorityArnList :: Lens' Tls (Maybe [Text])

List of ACM Certificate Authority ARNs.

Unauthenticated

unauthenticated_enabled :: Lens' Unauthenticated (Maybe Bool)

Specifies whether you want to enable or disable unauthenticated traffic to your cluster.

UnprocessedScramSecret

unprocessedScramSecret_errorCode :: Lens' UnprocessedScramSecret (Maybe Text)

Error code for associate/disassociate failure.

unprocessedScramSecret_errorMessage :: Lens' UnprocessedScramSecret (Maybe Text)

Error message for associate/disassociate failure.

unprocessedScramSecret_secretArn :: Lens' UnprocessedScramSecret (Maybe Text)

AWS Secrets Manager secret ARN.

ZookeeperNodeInfo

zookeeperNodeInfo_zookeeperVersion :: Lens' ZookeeperNodeInfo (Maybe Text)

The version of Zookeeper.

zookeeperNodeInfo_attachedENIId :: Lens' ZookeeperNodeInfo (Maybe Text)

The attached elastic network interface of the broker.

zookeeperNodeInfo_endpoints :: Lens' ZookeeperNodeInfo (Maybe [Text])

Endpoints for accessing the ZooKeeper.

zookeeperNodeInfo_clientVpcIpAddress :: Lens' ZookeeperNodeInfo (Maybe Text)

The virtual private cloud (VPC) IP address of the client.

zookeeperNodeInfo_zookeeperId :: Lens' ZookeeperNodeInfo (Maybe Double)

The role-specific ID for Zookeeper.