Network.AWS.Kafka.ListNodes

Description

Returns a list of the broker nodes in the cluster.

This operation returns paginated results.

Creating a Request

data ListNodes

See: newListNodes smart constructor.

Constructors

ListNodes'

Fields

nextToken :: Maybe Text

The paginated results marker. When the result of the operation is truncated, the call returns NextToken in the response. To get the next batch, provide this token in your next request.

maxResults :: Maybe Natural

The maximum number of results to return in the response. If there are more results, the response includes a NextToken parameter.

clusterArn :: Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

newListNodes

Arguments

:: Text
$sel:clusterArn:ListNodes'

-> ListNodes

Create a value of ListNodes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListNodes', listNodes_nextToken - The paginated results marker. When the result of the operation is truncated, the call returns NextToken in the response. To get the next batch, provide this token in your next request.

$sel:maxResults:ListNodes', listNodes_maxResults - The maximum number of results to return in the response. If there are more results, the response includes a NextToken parameter.

$sel:clusterArn:ListNodes', listNodes_clusterArn - The Amazon Resource Name (ARN) that uniquely identifies the cluster.

Request Lenses

listNodes_nextToken :: Lens' ListNodes (Maybe Text)

The paginated results marker. When the result of the operation is truncated, the call returns NextToken in the response. To get the next batch, provide this token in your next request.

listNodes_maxResults :: Lens' ListNodes (Maybe Natural)

The maximum number of results to return in the response. If there are more results, the response includes a NextToken parameter.

listNodes_clusterArn :: Lens' ListNodes Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

Destructuring the Response

data ListNodesResponse

See: newListNodesResponse smart constructor.

Constructors

ListNodesResponse'

Fields

nodeInfoList :: Maybe [NodeInfo]

List containing a NodeInfo object.

nextToken :: Maybe Text

The paginated results marker. When the result of a ListNodes operation is truncated, the call returns NextToken in the response. To get another batch of nodes, provide this token in your next request.

httpStatus :: Int

The response's http status code.

newListNodesResponse

Arguments

:: Int
$sel:httpStatus:ListNodesResponse'

-> ListNodesResponse

Create a value of ListNodesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nodeInfoList:ListNodesResponse', listNodesResponse_nodeInfoList - List containing a NodeInfo object.

$sel:nextToken:ListNodes', listNodesResponse_nextToken - The paginated results marker. When the result of a ListNodes operation is truncated, the call returns NextToken in the response. To get another batch of nodes, provide this token in your next request.

$sel:httpStatus:ListNodesResponse', listNodesResponse_httpStatus - The response's http status code.

Response Lenses

listNodesResponse_nodeInfoList :: Lens' ListNodesResponse (Maybe [NodeInfo])

List containing a NodeInfo object.

listNodesResponse_nextToken :: Lens' ListNodesResponse (Maybe Text)

The paginated results marker. When the result of a ListNodes operation is truncated, the call returns NextToken in the response. To get another batch of nodes, provide this token in your next request.

listNodesResponse_httpStatus :: Lens' ListNodesResponse Int

The response's http status code.