Network.AWS.Kafka.Types.BrokerNodeInfo

Description

Documentation

data BrokerNodeInfo

BrokerNodeInfo

See: newBrokerNodeInfo smart constructor.

Constructors

BrokerNodeInfo'

Fields

currentBrokerSoftwareInfo :: Maybe BrokerSoftwareInfo

Information about the version of software currently deployed on the Kafka brokers in the cluster.

clientSubnet :: Maybe Text

The client subnet to which this broker node belongs.

attachedENIId :: Maybe Text

The attached elastic network interface of the broker.

endpoints :: Maybe [Text]

Endpoints for accessing the broker.

clientVpcIpAddress :: Maybe Text

The virtual private cloud (VPC) of the client.

brokerId :: Maybe Double

The ID of the broker.

newBrokerNodeInfo :: BrokerNodeInfo

Create a value of BrokerNodeInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currentBrokerSoftwareInfo:BrokerNodeInfo', brokerNodeInfo_currentBrokerSoftwareInfo - Information about the version of software currently deployed on the Kafka brokers in the cluster.

$sel:clientSubnet:BrokerNodeInfo', brokerNodeInfo_clientSubnet - The client subnet to which this broker node belongs.

$sel:attachedENIId:BrokerNodeInfo', brokerNodeInfo_attachedENIId - The attached elastic network interface of the broker.

$sel:endpoints:BrokerNodeInfo', brokerNodeInfo_endpoints - Endpoints for accessing the broker.

$sel:clientVpcIpAddress:BrokerNodeInfo', brokerNodeInfo_clientVpcIpAddress - The virtual private cloud (VPC) of the client.

$sel:brokerId:BrokerNodeInfo', brokerNodeInfo_brokerId - The ID of the broker.

brokerNodeInfo_currentBrokerSoftwareInfo :: Lens' BrokerNodeInfo (Maybe BrokerSoftwareInfo)

Information about the version of software currently deployed on the Kafka brokers in the cluster.

brokerNodeInfo_clientSubnet :: Lens' BrokerNodeInfo (Maybe Text)

The client subnet to which this broker node belongs.

brokerNodeInfo_attachedENIId :: Lens' BrokerNodeInfo (Maybe Text)

The attached elastic network interface of the broker.

brokerNodeInfo_endpoints :: Lens' BrokerNodeInfo (Maybe [Text])

Endpoints for accessing the broker.

brokerNodeInfo_clientVpcIpAddress :: Lens' BrokerNodeInfo (Maybe Text)

The virtual private cloud (VPC) of the client.

brokerNodeInfo_brokerId :: Lens' BrokerNodeInfo (Maybe Double)

The ID of the broker.