Network.AWS.Kafka.Types.ClusterInfo

Description

Documentation

data ClusterInfo

Returns information about a cluster.

See: newClusterInfo smart constructor.

Constructors

ClusterInfo'

Fields

creationTime :: Maybe POSIX

The time when the cluster was created.

activeOperationArn :: Maybe Text

Arn of active cluster operation.

state :: Maybe ClusterState

The state of the cluster. The possible states are ACTIVE, CREATING, DELETING, FAILED, HEALING, MAINTENANCE, REBOOTING_BROKER, and UPDATING.

clusterArn :: Maybe Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

numberOfBrokerNodes :: Maybe Int

The number of broker nodes in the cluster.

enhancedMonitoring :: Maybe EnhancedMonitoring

Specifies which metrics are gathered for the MSK cluster. This property has the following possible values: DEFAULT, PER_BROKER, PER_TOPIC_PER_BROKER, and PER_TOPIC_PER_PARTITION. For a list of the metrics associated with each of these levels of monitoring, see Monitoring.

brokerNodeGroupInfo :: Maybe BrokerNodeGroupInfo

Information about the broker nodes.

openMonitoring :: Maybe OpenMonitoring

Settings for open monitoring using Prometheus.

currentBrokerSoftwareInfo :: Maybe BrokerSoftwareInfo

Information about the version of software currently deployed on the Kafka brokers in the cluster.

currentVersion :: Maybe Text

The current version of the MSK cluster.

stateInfo :: Maybe StateInfo

loggingInfo :: Maybe LoggingInfo

zookeeperConnectString :: Maybe Text

The connection string to use to connect to the Apache ZooKeeper cluster.

clusterName :: Maybe Text

The name of the cluster.

zookeeperConnectStringTls :: Maybe Text

The connection string to use to connect to zookeeper cluster on Tls port.

clientAuthentication :: Maybe ClientAuthentication

Includes all client authentication information.

tags :: Maybe (HashMap Text Text)

Tags attached to the cluster.

encryptionInfo :: Maybe EncryptionInfo

Includes all encryption-related information.

newClusterInfo :: ClusterInfo

Create a value of ClusterInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ClusterInfo', clusterInfo_creationTime - The time when the cluster was created.

$sel:activeOperationArn:ClusterInfo', clusterInfo_activeOperationArn - Arn of active cluster operation.

$sel:state:ClusterInfo', clusterInfo_state - The state of the cluster. The possible states are ACTIVE, CREATING, DELETING, FAILED, HEALING, MAINTENANCE, REBOOTING_BROKER, and UPDATING.

$sel:clusterArn:ClusterInfo', clusterInfo_clusterArn - The Amazon Resource Name (ARN) that uniquely identifies the cluster.

$sel:numberOfBrokerNodes:ClusterInfo', clusterInfo_numberOfBrokerNodes - The number of broker nodes in the cluster.

$sel:enhancedMonitoring:ClusterInfo', clusterInfo_enhancedMonitoring - Specifies which metrics are gathered for the MSK cluster. This property has the following possible values: DEFAULT, PER_BROKER, PER_TOPIC_PER_BROKER, and PER_TOPIC_PER_PARTITION. For a list of the metrics associated with each of these levels of monitoring, see Monitoring.

$sel:brokerNodeGroupInfo:ClusterInfo', clusterInfo_brokerNodeGroupInfo - Information about the broker nodes.

$sel:openMonitoring:ClusterInfo', clusterInfo_openMonitoring - Settings for open monitoring using Prometheus.

$sel:currentBrokerSoftwareInfo:ClusterInfo', clusterInfo_currentBrokerSoftwareInfo - Information about the version of software currently deployed on the Kafka brokers in the cluster.

$sel:currentVersion:ClusterInfo', clusterInfo_currentVersion - The current version of the MSK cluster.

$sel:stateInfo:ClusterInfo', clusterInfo_stateInfo - Undocumented member.

$sel:loggingInfo:ClusterInfo', clusterInfo_loggingInfo - Undocumented member.

$sel:zookeeperConnectString:ClusterInfo', clusterInfo_zookeeperConnectString - The connection string to use to connect to the Apache ZooKeeper cluster.

$sel:clusterName:ClusterInfo', clusterInfo_clusterName - The name of the cluster.

$sel:zookeeperConnectStringTls:ClusterInfo', clusterInfo_zookeeperConnectStringTls - The connection string to use to connect to zookeeper cluster on Tls port.

$sel:clientAuthentication:ClusterInfo', clusterInfo_clientAuthentication - Includes all client authentication information.

$sel:tags:ClusterInfo', clusterInfo_tags - Tags attached to the cluster.

$sel:encryptionInfo:ClusterInfo', clusterInfo_encryptionInfo - Includes all encryption-related information.

clusterInfo_creationTime :: Lens' ClusterInfo (Maybe UTCTime)

The time when the cluster was created.

clusterInfo_activeOperationArn :: Lens' ClusterInfo (Maybe Text)

Arn of active cluster operation.

clusterInfo_state :: Lens' ClusterInfo (Maybe ClusterState)

The state of the cluster. The possible states are ACTIVE, CREATING, DELETING, FAILED, HEALING, MAINTENANCE, REBOOTING_BROKER, and UPDATING.

clusterInfo_clusterArn :: Lens' ClusterInfo (Maybe Text)

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

clusterInfo_numberOfBrokerNodes :: Lens' ClusterInfo (Maybe Int)

The number of broker nodes in the cluster.

clusterInfo_enhancedMonitoring :: Lens' ClusterInfo (Maybe EnhancedMonitoring)

Specifies which metrics are gathered for the MSK cluster. This property has the following possible values: DEFAULT, PER_BROKER, PER_TOPIC_PER_BROKER, and PER_TOPIC_PER_PARTITION. For a list of the metrics associated with each of these levels of monitoring, see Monitoring.

clusterInfo_brokerNodeGroupInfo :: Lens' ClusterInfo (Maybe BrokerNodeGroupInfo)

Information about the broker nodes.

clusterInfo_openMonitoring :: Lens' ClusterInfo (Maybe OpenMonitoring)

Settings for open monitoring using Prometheus.

clusterInfo_currentBrokerSoftwareInfo :: Lens' ClusterInfo (Maybe BrokerSoftwareInfo)

Information about the version of software currently deployed on the Kafka brokers in the cluster.

clusterInfo_currentVersion :: Lens' ClusterInfo (Maybe Text)

The current version of the MSK cluster.

clusterInfo_stateInfo :: Lens' ClusterInfo (Maybe StateInfo)

Undocumented member.

clusterInfo_loggingInfo :: Lens' ClusterInfo (Maybe LoggingInfo)

Undocumented member.

clusterInfo_zookeeperConnectString :: Lens' ClusterInfo (Maybe Text)

The connection string to use to connect to the Apache ZooKeeper cluster.

clusterInfo_clusterName :: Lens' ClusterInfo (Maybe Text)

The name of the cluster.

clusterInfo_zookeeperConnectStringTls :: Lens' ClusterInfo (Maybe Text)

The connection string to use to connect to zookeeper cluster on Tls port.

clusterInfo_clientAuthentication :: Lens' ClusterInfo (Maybe ClientAuthentication)

Includes all client authentication information.

clusterInfo_tags :: Lens' ClusterInfo (Maybe (HashMap Text Text))

Tags attached to the cluster.

clusterInfo_encryptionInfo :: Lens' ClusterInfo (Maybe EncryptionInfo)

Includes all encryption-related information.