Network.AWS.Kafka.Types.Configuration

Description

Documentation

data Configuration

Represents an MSK Configuration.

See: newConfiguration smart constructor.

Constructors

Configuration'

Fields

description :: Text

The description of the configuration.

latestRevision :: ConfigurationRevision

Latest revision of the configuration.

creationTime :: POSIX

The time when the configuration was created.

kafkaVersions :: [Text]

An array of the versions of Apache Kafka with which you can use this MSK configuration. You can use this configuration for an MSK cluster only if the Apache Kafka version specified for the cluster appears in this array.

arn :: Text

The Amazon Resource Name (ARN) of the configuration.

name :: Text

The name of the configuration.

state :: ConfigurationState

The state of the configuration. The possible states are ACTIVE, DELETING, and DELETE_FAILED.

newConfiguration

Arguments

:: Text
$sel:description:Configuration'

-> ConfigurationRevision
$sel:latestRevision:Configuration'

-> UTCTime
$sel:creationTime:Configuration'

-> Text
$sel:arn:Configuration'

-> Text
$sel:name:Configuration'

-> ConfigurationState
$sel:state:Configuration'

-> Configuration

Create a value of Configuration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:Configuration', configuration_description - The description of the configuration.

$sel:latestRevision:Configuration', configuration_latestRevision - Latest revision of the configuration.

$sel:creationTime:Configuration', configuration_creationTime - The time when the configuration was created.

$sel:kafkaVersions:Configuration', configuration_kafkaVersions - An array of the versions of Apache Kafka with which you can use this MSK configuration. You can use this configuration for an MSK cluster only if the Apache Kafka version specified for the cluster appears in this array.

$sel:arn:Configuration', configuration_arn - The Amazon Resource Name (ARN) of the configuration.

$sel:name:Configuration', configuration_name - The name of the configuration.

$sel:state:Configuration', configuration_state - The state of the configuration. The possible states are ACTIVE, DELETING, and DELETE_FAILED.

configuration_description :: Lens' Configuration Text

The description of the configuration.

configuration_latestRevision :: Lens' Configuration ConfigurationRevision

Latest revision of the configuration.

configuration_creationTime :: Lens' Configuration UTCTime

The time when the configuration was created.

configuration_kafkaVersions :: Lens' Configuration [Text]

An array of the versions of Apache Kafka with which you can use this MSK configuration. You can use this configuration for an MSK cluster only if the Apache Kafka version specified for the cluster appears in this array.

configuration_arn :: Lens' Configuration Text

The Amazon Resource Name (ARN) of the configuration.

configuration_name :: Lens' Configuration Text

The name of the configuration.

configuration_state :: Lens' Configuration ConfigurationState

The state of the configuration. The possible states are ACTIVE, DELETING, and DELETE_FAILED.