Network.AWS.Kafka.Types.EncryptionInTransit

Description

Documentation

data EncryptionInTransit

The settings for encrypting data in transit.

See: newEncryptionInTransit smart constructor.

Constructors

EncryptionInTransit'

Fields

clientBroker :: Maybe ClientBroker

Indicates the encryption setting for data in transit between clients and brokers. The following are the possible values.

TLS means that client-broker communication is enabled with TLS only.

TLS_PLAINTEXT means that client-broker communication is enabled for both TLS-encrypted, as well as plaintext data.

PLAINTEXT means that client-broker communication is enabled in plaintext only.

The default value is TLS_PLAINTEXT.

inCluster :: Maybe Bool

When set to true, it indicates that data communication among the broker nodes of the cluster is encrypted. When set to false, the communication happens in plaintext.

The default value is true.

newEncryptionInTransit :: EncryptionInTransit

Create a value of EncryptionInTransit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientBroker:EncryptionInTransit', encryptionInTransit_clientBroker - Indicates the encryption setting for data in transit between clients and brokers. The following are the possible values.

TLS means that client-broker communication is enabled with TLS only.

TLS_PLAINTEXT means that client-broker communication is enabled for both TLS-encrypted, as well as plaintext data.

PLAINTEXT means that client-broker communication is enabled in plaintext only.

The default value is TLS_PLAINTEXT.

$sel:inCluster:EncryptionInTransit', encryptionInTransit_inCluster - When set to true, it indicates that data communication among the broker nodes of the cluster is encrypted. When set to false, the communication happens in plaintext.

The default value is true.

encryptionInTransit_clientBroker :: Lens' EncryptionInTransit (Maybe ClientBroker)

Indicates the encryption setting for data in transit between clients and brokers. The following are the possible values.

TLS means that client-broker communication is enabled with TLS only.

TLS_PLAINTEXT means that client-broker communication is enabled for both TLS-encrypted, as well as plaintext data.

PLAINTEXT means that client-broker communication is enabled in plaintext only.

The default value is TLS_PLAINTEXT.

encryptionInTransit_inCluster :: Lens' EncryptionInTransit (Maybe Bool)

When set to true, it indicates that data communication among the broker nodes of the cluster is encrypted. When set to false, the communication happens in plaintext.

The default value is true.