Network.AWS.Kafka.Types.MutableClusterInfo

Description

Documentation

data MutableClusterInfo

Information about cluster attributes that can be updated via update APIs.

See: newMutableClusterInfo smart constructor.

Constructors

MutableClusterInfo'

Fields

numberOfBrokerNodes :: Maybe Int

The number of broker nodes in the cluster.

enhancedMonitoring :: Maybe EnhancedMonitoring

Specifies which Apache Kafka metrics Amazon MSK gathers and sends to Amazon CloudWatch for this cluster.

openMonitoring :: Maybe OpenMonitoring

The settings for open monitoring.

configurationInfo :: Maybe ConfigurationInfo

Information about the changes in the configuration of the brokers.

instanceType :: Maybe Text

Information about the Amazon MSK broker type.

kafkaVersion :: Maybe Text

The Kafka version.

loggingInfo :: Maybe LoggingInfo

You can configure your MSK cluster to send broker logs to different destination types. This is a container for the configuration details related to broker logs.

clientAuthentication :: Maybe ClientAuthentication

Includes all client authentication information.

brokerEBSVolumeInfo :: Maybe [BrokerEBSVolumeInfo]

Specifies the size of the EBS volume and the ID of the associated broker.

encryptionInfo :: Maybe EncryptionInfo

Includes all encryption-related information.

newMutableClusterInfo :: MutableClusterInfo

Create a value of MutableClusterInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:numberOfBrokerNodes:MutableClusterInfo', mutableClusterInfo_numberOfBrokerNodes - The number of broker nodes in the cluster.

$sel:enhancedMonitoring:MutableClusterInfo', mutableClusterInfo_enhancedMonitoring - Specifies which Apache Kafka metrics Amazon MSK gathers and sends to Amazon CloudWatch for this cluster.

$sel:openMonitoring:MutableClusterInfo', mutableClusterInfo_openMonitoring - The settings for open monitoring.

$sel:configurationInfo:MutableClusterInfo', mutableClusterInfo_configurationInfo - Information about the changes in the configuration of the brokers.

$sel:instanceType:MutableClusterInfo', mutableClusterInfo_instanceType - Information about the Amazon MSK broker type.

$sel:kafkaVersion:MutableClusterInfo', mutableClusterInfo_kafkaVersion - The Kafka version.

$sel:loggingInfo:MutableClusterInfo', mutableClusterInfo_loggingInfo - You can configure your MSK cluster to send broker logs to different destination types. This is a container for the configuration details related to broker logs.

$sel:clientAuthentication:MutableClusterInfo', mutableClusterInfo_clientAuthentication - Includes all client authentication information.

$sel:brokerEBSVolumeInfo:MutableClusterInfo', mutableClusterInfo_brokerEBSVolumeInfo - Specifies the size of the EBS volume and the ID of the associated broker.

$sel:encryptionInfo:MutableClusterInfo', mutableClusterInfo_encryptionInfo - Includes all encryption-related information.

mutableClusterInfo_numberOfBrokerNodes :: Lens' MutableClusterInfo (Maybe Int)

The number of broker nodes in the cluster.

mutableClusterInfo_enhancedMonitoring :: Lens' MutableClusterInfo (Maybe EnhancedMonitoring)

Specifies which Apache Kafka metrics Amazon MSK gathers and sends to Amazon CloudWatch for this cluster.

mutableClusterInfo_openMonitoring :: Lens' MutableClusterInfo (Maybe OpenMonitoring)

The settings for open monitoring.

mutableClusterInfo_configurationInfo :: Lens' MutableClusterInfo (Maybe ConfigurationInfo)

Information about the changes in the configuration of the brokers.

mutableClusterInfo_instanceType :: Lens' MutableClusterInfo (Maybe Text)

Information about the Amazon MSK broker type.

mutableClusterInfo_kafkaVersion :: Lens' MutableClusterInfo (Maybe Text)

The Kafka version.

mutableClusterInfo_loggingInfo :: Lens' MutableClusterInfo (Maybe LoggingInfo)

You can configure your MSK cluster to send broker logs to different destination types. This is a container for the configuration details related to broker logs.

mutableClusterInfo_clientAuthentication :: Lens' MutableClusterInfo (Maybe ClientAuthentication)

Includes all client authentication information.

mutableClusterInfo_brokerEBSVolumeInfo :: Lens' MutableClusterInfo (Maybe [BrokerEBSVolumeInfo])

Specifies the size of the EBS volume and the ID of the associated broker.

mutableClusterInfo_encryptionInfo :: Lens' MutableClusterInfo (Maybe EncryptionInfo)

Includes all encryption-related information.