Network.AWS.Kafka.Types.NodeExporterInfo

Description

Documentation

data NodeExporterInfo

Indicates whether you want to enable or disable the Node Exporter.

See: newNodeExporterInfo smart constructor.

Constructors

NodeExporterInfo'

Fields

enabledInBroker :: Bool

Indicates whether you want to enable or disable the Node Exporter.

newNodeExporterInfo

Arguments

:: Bool
$sel:enabledInBroker:NodeExporterInfo'

-> NodeExporterInfo

Create a value of NodeExporterInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabledInBroker:NodeExporterInfo', nodeExporterInfo_enabledInBroker - Indicates whether you want to enable or disable the Node Exporter.

nodeExporterInfo_enabledInBroker :: Lens' NodeExporterInfo Bool

Indicates whether you want to enable or disable the Node Exporter.