Network.AWS.Kafka.UpdateClusterConfiguration

Description

Updates the cluster with the configuration that is specified in the request body.

Creating a Request

data UpdateClusterConfiguration

See: newUpdateClusterConfiguration smart constructor.

Constructors

UpdateClusterConfiguration'

Fields

clusterArn :: Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

currentVersion :: Text

The version of the cluster that needs to be updated.

configurationInfo :: ConfigurationInfo

Represents the configuration that you want MSK to use for the brokers in a cluster.

newUpdateClusterConfiguration

Arguments

:: Text
$sel:clusterArn:UpdateClusterConfiguration'

-> Text
$sel:currentVersion:UpdateClusterConfiguration'

-> ConfigurationInfo
$sel:configurationInfo:UpdateClusterConfiguration'

-> UpdateClusterConfiguration

Create a value of UpdateClusterConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterArn:UpdateClusterConfiguration', updateClusterConfiguration_clusterArn - The Amazon Resource Name (ARN) that uniquely identifies the cluster.

$sel:currentVersion:UpdateClusterConfiguration', updateClusterConfiguration_currentVersion - The version of the cluster that needs to be updated.

$sel:configurationInfo:UpdateClusterConfiguration', updateClusterConfiguration_configurationInfo - Represents the configuration that you want MSK to use for the brokers in a cluster.

Request Lenses

updateClusterConfiguration_clusterArn :: Lens' UpdateClusterConfiguration Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

updateClusterConfiguration_currentVersion :: Lens' UpdateClusterConfiguration Text

The version of the cluster that needs to be updated.

updateClusterConfiguration_configurationInfo :: Lens' UpdateClusterConfiguration ConfigurationInfo

Represents the configuration that you want MSK to use for the brokers in a cluster.

Destructuring the Response

data UpdateClusterConfigurationResponse

See: newUpdateClusterConfigurationResponse smart constructor.

Constructors

UpdateClusterConfigurationResponse'

Fields

clusterArn :: Maybe Text

The Amazon Resource Name (ARN) of the cluster.

clusterOperationArn :: Maybe Text

The Amazon Resource Name (ARN) of the cluster operation.

httpStatus :: Int

The response's http status code.

newUpdateClusterConfigurationResponse

Arguments

:: Int
$sel:httpStatus:UpdateClusterConfigurationResponse'

-> UpdateClusterConfigurationResponse

Create a value of UpdateClusterConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterArn:UpdateClusterConfiguration', updateClusterConfigurationResponse_clusterArn - The Amazon Resource Name (ARN) of the cluster.

$sel:clusterOperationArn:UpdateClusterConfigurationResponse', updateClusterConfigurationResponse_clusterOperationArn - The Amazon Resource Name (ARN) of the cluster operation.

$sel:httpStatus:UpdateClusterConfigurationResponse', updateClusterConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

updateClusterConfigurationResponse_clusterArn :: Lens' UpdateClusterConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

updateClusterConfigurationResponse_clusterOperationArn :: Lens' UpdateClusterConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster operation.

updateClusterConfigurationResponse_httpStatus :: Lens' UpdateClusterConfigurationResponse Int

The response's http status code.