Network.AWS.Kafka.UpdateMonitoring

Description

Updates the monitoring settings for the cluster. You can use this operation to specify which Apache Kafka metrics you want Amazon MSK to send to Amazon CloudWatch. You can also specify settings for open monitoring with Prometheus.

Creating a Request

data UpdateMonitoring

Request body for UpdateMonitoring.

See: newUpdateMonitoring smart constructor.

Constructors

UpdateMonitoring'

Fields

enhancedMonitoring :: Maybe EnhancedMonitoring

Specifies which Apache Kafka metrics Amazon MSK gathers and sends to Amazon CloudWatch for this cluster.

openMonitoring :: Maybe OpenMonitoringInfo

The settings for open monitoring.

loggingInfo :: Maybe LoggingInfo

clusterArn :: Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

currentVersion :: Text

The version of the MSK cluster to update. Cluster versions aren't simple numbers. You can describe an MSK cluster to find its version. When this update operation is successful, it generates a new cluster version.

newUpdateMonitoring

Arguments

:: Text
$sel:clusterArn:UpdateMonitoring'

-> Text
$sel:currentVersion:UpdateMonitoring'

-> UpdateMonitoring

Create a value of UpdateMonitoring with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enhancedMonitoring:UpdateMonitoring', updateMonitoring_enhancedMonitoring - Specifies which Apache Kafka metrics Amazon MSK gathers and sends to Amazon CloudWatch for this cluster.

$sel:openMonitoring:UpdateMonitoring', updateMonitoring_openMonitoring - The settings for open monitoring.

$sel:loggingInfo:UpdateMonitoring', updateMonitoring_loggingInfo - Undocumented member.

$sel:clusterArn:UpdateMonitoring', updateMonitoring_clusterArn - The Amazon Resource Name (ARN) that uniquely identifies the cluster.

$sel:currentVersion:UpdateMonitoring', updateMonitoring_currentVersion - The version of the MSK cluster to update. Cluster versions aren't simple numbers. You can describe an MSK cluster to find its version. When this update operation is successful, it generates a new cluster version.

Request Lenses

updateMonitoring_enhancedMonitoring :: Lens' UpdateMonitoring (Maybe EnhancedMonitoring)

Specifies which Apache Kafka metrics Amazon MSK gathers and sends to Amazon CloudWatch for this cluster.

updateMonitoring_openMonitoring :: Lens' UpdateMonitoring (Maybe OpenMonitoringInfo)

The settings for open monitoring.

updateMonitoring_loggingInfo :: Lens' UpdateMonitoring (Maybe LoggingInfo)

Undocumented member.

updateMonitoring_clusterArn :: Lens' UpdateMonitoring Text

The Amazon Resource Name (ARN) that uniquely identifies the cluster.

updateMonitoring_currentVersion :: Lens' UpdateMonitoring Text

The version of the MSK cluster to update. Cluster versions aren't simple numbers. You can describe an MSK cluster to find its version. When this update operation is successful, it generates a new cluster version.

Destructuring the Response

data UpdateMonitoringResponse

See: newUpdateMonitoringResponse smart constructor.

Constructors

UpdateMonitoringResponse'

Fields

clusterArn :: Maybe Text

The Amazon Resource Name (ARN) of the cluster.

clusterOperationArn :: Maybe Text

The Amazon Resource Name (ARN) of the cluster operation.

httpStatus :: Int

The response's http status code.

newUpdateMonitoringResponse

Arguments

:: Int
$sel:httpStatus:UpdateMonitoringResponse'

-> UpdateMonitoringResponse

Create a value of UpdateMonitoringResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterArn:UpdateMonitoring', updateMonitoringResponse_clusterArn - The Amazon Resource Name (ARN) of the cluster.

$sel:clusterOperationArn:UpdateMonitoringResponse', updateMonitoringResponse_clusterOperationArn - The Amazon Resource Name (ARN) of the cluster operation.

$sel:httpStatus:UpdateMonitoringResponse', updateMonitoringResponse_httpStatus - The response's http status code.

Response Lenses

updateMonitoringResponse_clusterArn :: Lens' UpdateMonitoringResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster.

updateMonitoringResponse_clusterOperationArn :: Lens' UpdateMonitoringResponse (Maybe Text)

The Amazon Resource Name (ARN) of the cluster operation.

updateMonitoringResponse_httpStatus :: Lens' UpdateMonitoringResponse Int

The response's http status code.