Network.AWS.KafkaConnect.CreateConnector

Description

Creates a connector using the specified properties.

Creating a Request

data CreateConnector

See: newCreateConnector smart constructor.

Constructors

CreateConnector'

Fields

logDelivery :: Maybe LogDelivery

Details about log delivery.

workerConfiguration :: Maybe WorkerConfiguration

Specifies which worker configuration to use with the connector.

connectorDescription :: Maybe Text

A summary description of the connector.

capacity :: Capacity

Information about the capacity allocated to the connector. Exactly one of the two properties must be specified.

connectorConfiguration :: HashMap Text Text

A map of keys to values that represent the configuration for the connector.

connectorName :: Text

The name of the connector.

kafkaCluster :: KafkaCluster

Specifies which Apache Kafka cluster to connect to.

kafkaClusterClientAuthentication :: KafkaClusterClientAuthentication

Details of the client authentication used by the Apache Kafka cluster.

kafkaClusterEncryptionInTransit :: KafkaClusterEncryptionInTransit

Details of encryption in transit to the Apache Kafka cluster.

kafkaConnectVersion :: Text

The version of Kafka Connect. It has to be compatible with both the Apache Kafka cluster's version and the plugins.

plugins :: [Plugin]

Specifies which plugins to use for the connector.

serviceExecutionRoleArn :: Text

The Amazon Resource Name (ARN) of the IAM role used by the connector to access the Amazon Web Services resources that it needs. The types of resources depends on the logic of the connector. For example, a connector that has Amazon S3 as a destination must have permissions that allow it to write to the S3 destination bucket.

newCreateConnector

Arguments

:: Capacity
$sel:capacity:CreateConnector'

-> Text
$sel:connectorName:CreateConnector'

-> KafkaCluster
$sel:kafkaCluster:CreateConnector'

-> KafkaClusterClientAuthentication
$sel:kafkaClusterClientAuthentication:CreateConnector'

-> KafkaClusterEncryptionInTransit
$sel:kafkaClusterEncryptionInTransit:CreateConnector'

-> Text
$sel:kafkaConnectVersion:CreateConnector'

-> Text
$sel:serviceExecutionRoleArn:CreateConnector'

-> CreateConnector

Create a value of CreateConnector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logDelivery:CreateConnector', createConnector_logDelivery - Details about log delivery.

$sel:workerConfiguration:CreateConnector', createConnector_workerConfiguration - Specifies which worker configuration to use with the connector.

$sel:connectorDescription:CreateConnector', createConnector_connectorDescription - A summary description of the connector.

$sel:capacity:CreateConnector', createConnector_capacity - Information about the capacity allocated to the connector. Exactly one of the two properties must be specified.

$sel:connectorConfiguration:CreateConnector', createConnector_connectorConfiguration - A map of keys to values that represent the configuration for the connector.

$sel:connectorName:CreateConnector', createConnector_connectorName - The name of the connector.

$sel:kafkaCluster:CreateConnector', createConnector_kafkaCluster - Specifies which Apache Kafka cluster to connect to.

$sel:kafkaClusterClientAuthentication:CreateConnector', createConnector_kafkaClusterClientAuthentication - Details of the client authentication used by the Apache Kafka cluster.

$sel:kafkaClusterEncryptionInTransit:CreateConnector', createConnector_kafkaClusterEncryptionInTransit - Details of encryption in transit to the Apache Kafka cluster.

$sel:kafkaConnectVersion:CreateConnector', createConnector_kafkaConnectVersion - The version of Kafka Connect. It has to be compatible with both the Apache Kafka cluster's version and the plugins.

$sel:plugins:CreateConnector', createConnector_plugins - Specifies which plugins to use for the connector.

$sel:serviceExecutionRoleArn:CreateConnector', createConnector_serviceExecutionRoleArn - The Amazon Resource Name (ARN) of the IAM role used by the connector to access the Amazon Web Services resources that it needs. The types of resources depends on the logic of the connector. For example, a connector that has Amazon S3 as a destination must have permissions that allow it to write to the S3 destination bucket.

Request Lenses

createConnector_logDelivery :: Lens' CreateConnector (Maybe LogDelivery)

Details about log delivery.

createConnector_workerConfiguration :: Lens' CreateConnector (Maybe WorkerConfiguration)

Specifies which worker configuration to use with the connector.

createConnector_connectorDescription :: Lens' CreateConnector (Maybe Text)

A summary description of the connector.

createConnector_capacity :: Lens' CreateConnector Capacity

Information about the capacity allocated to the connector. Exactly one of the two properties must be specified.

createConnector_connectorConfiguration :: Lens' CreateConnector (HashMap Text Text)

A map of keys to values that represent the configuration for the connector.

createConnector_connectorName :: Lens' CreateConnector Text

The name of the connector.

createConnector_kafkaCluster :: Lens' CreateConnector KafkaCluster

Specifies which Apache Kafka cluster to connect to.

createConnector_kafkaClusterClientAuthentication :: Lens' CreateConnector KafkaClusterClientAuthentication

Details of the client authentication used by the Apache Kafka cluster.

createConnector_kafkaClusterEncryptionInTransit :: Lens' CreateConnector KafkaClusterEncryptionInTransit

Details of encryption in transit to the Apache Kafka cluster.

createConnector_kafkaConnectVersion :: Lens' CreateConnector Text

The version of Kafka Connect. It has to be compatible with both the Apache Kafka cluster's version and the plugins.

createConnector_plugins :: Lens' CreateConnector [Plugin]

Specifies which plugins to use for the connector.

createConnector_serviceExecutionRoleArn :: Lens' CreateConnector Text

The Amazon Resource Name (ARN) of the IAM role used by the connector to access the Amazon Web Services resources that it needs. The types of resources depends on the logic of the connector. For example, a connector that has Amazon S3 as a destination must have permissions that allow it to write to the S3 destination bucket.

Destructuring the Response

data CreateConnectorResponse

See: newCreateConnectorResponse smart constructor.

Constructors

CreateConnectorResponse'

Fields

connectorArn :: Maybe Text

The Amazon Resource Name (ARN) that Amazon assigned to the connector.

connectorName :: Maybe Text

The name of the connector.

connectorState :: Maybe ConnectorState

The state of the connector.

httpStatus :: Int

The response's http status code.

newCreateConnectorResponse

Arguments

:: Int
$sel:httpStatus:CreateConnectorResponse'

-> CreateConnectorResponse

Create a value of CreateConnectorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectorArn:CreateConnectorResponse', createConnectorResponse_connectorArn - The Amazon Resource Name (ARN) that Amazon assigned to the connector.

$sel:connectorName:CreateConnector', createConnectorResponse_connectorName - The name of the connector.

$sel:connectorState:CreateConnectorResponse', createConnectorResponse_connectorState - The state of the connector.

$sel:httpStatus:CreateConnectorResponse', createConnectorResponse_httpStatus - The response's http status code.

Response Lenses

createConnectorResponse_connectorArn :: Lens' CreateConnectorResponse (Maybe Text)

The Amazon Resource Name (ARN) that Amazon assigned to the connector.

createConnectorResponse_connectorName :: Lens' CreateConnectorResponse (Maybe Text)

The name of the connector.

createConnectorResponse_connectorState :: Lens' CreateConnectorResponse (Maybe ConnectorState)

The state of the connector.

createConnectorResponse_httpStatus :: Lens' CreateConnectorResponse Int

The response's http status code.