Network.AWS.KafkaConnect.CreateWorkerConfiguration

Description

Creates a worker configuration using the specified properties.

Creating a Request

data CreateWorkerConfiguration

See: newCreateWorkerConfiguration smart constructor.

Constructors

CreateWorkerConfiguration'

Fields

description :: Maybe Text

A summary description of the worker configuration.

name :: Text

The name of the worker configuration.

propertiesFileContent :: Text

Base64 encoded contents of connect-distributed.properties file.

newCreateWorkerConfiguration

Arguments

:: Text
$sel:name:CreateWorkerConfiguration'

-> Text
$sel:propertiesFileContent:CreateWorkerConfiguration'

-> CreateWorkerConfiguration

Create a value of CreateWorkerConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateWorkerConfiguration', createWorkerConfiguration_description - A summary description of the worker configuration.

$sel:name:CreateWorkerConfiguration', createWorkerConfiguration_name - The name of the worker configuration.

$sel:propertiesFileContent:CreateWorkerConfiguration', createWorkerConfiguration_propertiesFileContent - Base64 encoded contents of connect-distributed.properties file.

Request Lenses

createWorkerConfiguration_description :: Lens' CreateWorkerConfiguration (Maybe Text)

A summary description of the worker configuration.

createWorkerConfiguration_name :: Lens' CreateWorkerConfiguration Text

The name of the worker configuration.

createWorkerConfiguration_propertiesFileContent :: Lens' CreateWorkerConfiguration Text

Base64 encoded contents of connect-distributed.properties file.

Destructuring the Response

data CreateWorkerConfigurationResponse

See: newCreateWorkerConfigurationResponse smart constructor.

Constructors

CreateWorkerConfigurationResponse'

Fields

creationTime :: Maybe POSIX

The time that the worker configuration was created.

latestRevision :: Maybe WorkerConfigurationRevisionSummary

The latest revision of the worker configuration.

name :: Maybe Text

The name of the worker configuration.

workerConfigurationArn :: Maybe Text

The Amazon Resource Name (ARN) that Amazon assigned to the worker configuration.

httpStatus :: Int

The response's http status code.

newCreateWorkerConfigurationResponse

Arguments

:: Int
$sel:httpStatus:CreateWorkerConfigurationResponse'

-> CreateWorkerConfigurationResponse

Create a value of CreateWorkerConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:CreateWorkerConfigurationResponse', createWorkerConfigurationResponse_creationTime - The time that the worker configuration was created.

$sel:latestRevision:CreateWorkerConfigurationResponse', createWorkerConfigurationResponse_latestRevision - The latest revision of the worker configuration.

$sel:name:CreateWorkerConfiguration', createWorkerConfigurationResponse_name - The name of the worker configuration.

$sel:workerConfigurationArn:CreateWorkerConfigurationResponse', createWorkerConfigurationResponse_workerConfigurationArn - The Amazon Resource Name (ARN) that Amazon assigned to the worker configuration.

$sel:httpStatus:CreateWorkerConfigurationResponse', createWorkerConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

createWorkerConfigurationResponse_creationTime :: Lens' CreateWorkerConfigurationResponse (Maybe UTCTime)

The time that the worker configuration was created.

createWorkerConfigurationResponse_latestRevision :: Lens' CreateWorkerConfigurationResponse (Maybe WorkerConfigurationRevisionSummary)

The latest revision of the worker configuration.

createWorkerConfigurationResponse_name :: Lens' CreateWorkerConfigurationResponse (Maybe Text)

The name of the worker configuration.

createWorkerConfigurationResponse_workerConfigurationArn :: Lens' CreateWorkerConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) that Amazon assigned to the worker configuration.

createWorkerConfigurationResponse_httpStatus :: Lens' CreateWorkerConfigurationResponse Int

The response's http status code.