Network.AWS.KafkaConnect.DescribeCustomPlugin

Description

A summary description of the custom plugin.

Creating a Request

data DescribeCustomPlugin

See: newDescribeCustomPlugin smart constructor.

Constructors

DescribeCustomPlugin'

Fields

customPluginArn :: Text

Returns information about a custom plugin.

newDescribeCustomPlugin

Arguments

:: Text
$sel:customPluginArn:DescribeCustomPlugin'

-> DescribeCustomPlugin

Create a value of DescribeCustomPlugin with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customPluginArn:DescribeCustomPlugin', describeCustomPlugin_customPluginArn - Returns information about a custom plugin.

Request Lenses

describeCustomPlugin_customPluginArn :: Lens' DescribeCustomPlugin Text

Returns information about a custom plugin.

Destructuring the Response

data DescribeCustomPluginResponse

See: newDescribeCustomPluginResponse smart constructor.

Constructors

DescribeCustomPluginResponse'

Fields

creationTime :: Maybe POSIX

The time that the custom plugin was created.

latestRevision :: Maybe CustomPluginRevisionSummary

The latest successfully created revision of the custom plugin. If there are no successfully created revisions, this field will be absent.

name :: Maybe Text

The name of the custom plugin.

customPluginArn :: Maybe Text

The Amazon Resource Name (ARN) of the custom plugin.

customPluginState :: Maybe CustomPluginState

The state of the custom plugin.

description :: Maybe Text

The description of the custom plugin.

httpStatus :: Int

The response's http status code.

newDescribeCustomPluginResponse

Arguments

:: Int
$sel:httpStatus:DescribeCustomPluginResponse'

-> DescribeCustomPluginResponse

Create a value of DescribeCustomPluginResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeCustomPluginResponse', describeCustomPluginResponse_creationTime - The time that the custom plugin was created.

$sel:latestRevision:DescribeCustomPluginResponse', describeCustomPluginResponse_latestRevision - The latest successfully created revision of the custom plugin. If there are no successfully created revisions, this field will be absent.

$sel:name:DescribeCustomPluginResponse', describeCustomPluginResponse_name - The name of the custom plugin.

$sel:customPluginArn:DescribeCustomPlugin', describeCustomPluginResponse_customPluginArn - The Amazon Resource Name (ARN) of the custom plugin.

$sel:customPluginState:DescribeCustomPluginResponse', describeCustomPluginResponse_customPluginState - The state of the custom plugin.

$sel:description:DescribeCustomPluginResponse', describeCustomPluginResponse_description - The description of the custom plugin.

$sel:httpStatus:DescribeCustomPluginResponse', describeCustomPluginResponse_httpStatus - The response's http status code.

Response Lenses

describeCustomPluginResponse_creationTime :: Lens' DescribeCustomPluginResponse (Maybe UTCTime)

The time that the custom plugin was created.

describeCustomPluginResponse_latestRevision :: Lens' DescribeCustomPluginResponse (Maybe CustomPluginRevisionSummary)

The latest successfully created revision of the custom plugin. If there are no successfully created revisions, this field will be absent.

describeCustomPluginResponse_name :: Lens' DescribeCustomPluginResponse (Maybe Text)

The name of the custom plugin.

describeCustomPluginResponse_customPluginArn :: Lens' DescribeCustomPluginResponse (Maybe Text)

The Amazon Resource Name (ARN) of the custom plugin.

describeCustomPluginResponse_customPluginState :: Lens' DescribeCustomPluginResponse (Maybe CustomPluginState)

The state of the custom plugin.

describeCustomPluginResponse_description :: Lens' DescribeCustomPluginResponse (Maybe Text)

The description of the custom plugin.

describeCustomPluginResponse_httpStatus :: Lens' DescribeCustomPluginResponse Int

The response's http status code.