Network.AWS.KafkaConnect.Lens

Description

Operations

ListWorkerConfigurations

listWorkerConfigurations_nextToken :: Lens' ListWorkerConfigurations (Maybe Text)

If the response of a ListWorkerConfigurations operation is truncated, it will include a NextToken. Send this NextToken in a subsequent request to continue listing from where the previous operation left off.

listWorkerConfigurations_maxResults :: Lens' ListWorkerConfigurations (Maybe Natural)

The maximum number of worker configurations to list in one response.

listWorkerConfigurationsResponse_workerConfigurations :: Lens' ListWorkerConfigurationsResponse (Maybe [WorkerConfigurationSummary])

An array of worker configuration descriptions.

listWorkerConfigurationsResponse_nextToken :: Lens' ListWorkerConfigurationsResponse (Maybe Text)

If the response of a ListWorkerConfigurations operation is truncated, it will include a NextToken. Send this NextToken in a subsequent request to continue listing from where the previous operation left off.

listWorkerConfigurationsResponse_httpStatus :: Lens' ListWorkerConfigurationsResponse Int

The response's http status code.

DescribeCustomPlugin

describeCustomPlugin_customPluginArn :: Lens' DescribeCustomPlugin Text

Returns information about a custom plugin.

describeCustomPluginResponse_creationTime :: Lens' DescribeCustomPluginResponse (Maybe UTCTime)

The time that the custom plugin was created.

describeCustomPluginResponse_latestRevision :: Lens' DescribeCustomPluginResponse (Maybe CustomPluginRevisionSummary)

The latest successfully created revision of the custom plugin. If there are no successfully created revisions, this field will be absent.

describeCustomPluginResponse_name :: Lens' DescribeCustomPluginResponse (Maybe Text)

The name of the custom plugin.

describeCustomPluginResponse_customPluginArn :: Lens' DescribeCustomPluginResponse (Maybe Text)

The Amazon Resource Name (ARN) of the custom plugin.

describeCustomPluginResponse_customPluginState :: Lens' DescribeCustomPluginResponse (Maybe CustomPluginState)

The state of the custom plugin.

describeCustomPluginResponse_description :: Lens' DescribeCustomPluginResponse (Maybe Text)

The description of the custom plugin.

describeCustomPluginResponse_httpStatus :: Lens' DescribeCustomPluginResponse Int

The response's http status code.

DeleteConnector

deleteConnector_currentVersion :: Lens' DeleteConnector (Maybe Text)

The current version of the connector that you want to delete.

deleteConnector_connectorArn :: Lens' DeleteConnector Text

The Amazon Resource Name (ARN) of the connector that you want to delete.

deleteConnectorResponse_connectorArn :: Lens' DeleteConnectorResponse (Maybe Text)

The Amazon Resource Name (ARN) of the connector that you requested to delete.

deleteConnectorResponse_connectorState :: Lens' DeleteConnectorResponse (Maybe ConnectorState)

The state of the connector that you requested to delete.

deleteConnectorResponse_httpStatus :: Lens' DeleteConnectorResponse Int

The response's http status code.

UpdateConnector

updateConnector_capacity :: Lens' UpdateConnector CapacityUpdate

The target capacity.

updateConnector_connectorArn :: Lens' UpdateConnector Text

The Amazon Resource Name (ARN) of the connector that you want to update.

updateConnector_currentVersion :: Lens' UpdateConnector Text

The current version of the connector that you want to update.

updateConnectorResponse_connectorArn :: Lens' UpdateConnectorResponse (Maybe Text)

The Amazon Resource Name (ARN) of the connector.

updateConnectorResponse_connectorState :: Lens' UpdateConnectorResponse (Maybe ConnectorState)

The state of the connector.

updateConnectorResponse_httpStatus :: Lens' UpdateConnectorResponse Int

The response's http status code.

CreateWorkerConfiguration

createWorkerConfiguration_description :: Lens' CreateWorkerConfiguration (Maybe Text)

A summary description of the worker configuration.

createWorkerConfiguration_name :: Lens' CreateWorkerConfiguration Text

The name of the worker configuration.

createWorkerConfiguration_propertiesFileContent :: Lens' CreateWorkerConfiguration Text

Base64 encoded contents of connect-distributed.properties file.

createWorkerConfigurationResponse_creationTime :: Lens' CreateWorkerConfigurationResponse (Maybe UTCTime)

The time that the worker configuration was created.

createWorkerConfigurationResponse_latestRevision :: Lens' CreateWorkerConfigurationResponse (Maybe WorkerConfigurationRevisionSummary)

The latest revision of the worker configuration.

createWorkerConfigurationResponse_name :: Lens' CreateWorkerConfigurationResponse (Maybe Text)

The name of the worker configuration.

createWorkerConfigurationResponse_workerConfigurationArn :: Lens' CreateWorkerConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) that Amazon assigned to the worker configuration.

createWorkerConfigurationResponse_httpStatus :: Lens' CreateWorkerConfigurationResponse Int

The response's http status code.

ListConnectors

listConnectors_connectorNamePrefix :: Lens' ListConnectors (Maybe Text)

The name prefix that you want to use to search for and list connectors.

listConnectors_nextToken :: Lens' ListConnectors (Maybe Text)

If the response of a ListConnectors operation is truncated, it will include a NextToken. Send this NextToken in a subsequent request to continue listing from where the previous operation left off.

listConnectors_maxResults :: Lens' ListConnectors (Maybe Natural)

The maximum number of connectors to list in one response.

listConnectorsResponse_connectors :: Lens' ListConnectorsResponse (Maybe [ConnectorSummary])

An array of connector descriptions.

listConnectorsResponse_nextToken :: Lens' ListConnectorsResponse (Maybe Text)

If the response of a ListConnectors operation is truncated, it will include a NextToken. Send this NextToken in a subsequent request to continue listing from where it left off.

listConnectorsResponse_httpStatus :: Lens' ListConnectorsResponse Int

The response's http status code.

ListCustomPlugins

listCustomPlugins_nextToken :: Lens' ListCustomPlugins (Maybe Text)

If the response of a ListCustomPlugins operation is truncated, it will include a NextToken. Send this NextToken in a subsequent request to continue listing from where the previous operation left off.

listCustomPlugins_maxResults :: Lens' ListCustomPlugins (Maybe Natural)

The maximum number of custom plugins to list in one response.

listCustomPluginsResponse_customPlugins :: Lens' ListCustomPluginsResponse (Maybe [CustomPluginSummary])

An array of custom plugin descriptions.

listCustomPluginsResponse_nextToken :: Lens' ListCustomPluginsResponse (Maybe Text)

If the response of a ListCustomPlugins operation is truncated, it will include a NextToken. Send this NextToken in a subsequent request to continue listing from where the previous operation left off.

listCustomPluginsResponse_httpStatus :: Lens' ListCustomPluginsResponse Int

The response's http status code.

CreateConnector

createConnector_logDelivery :: Lens' CreateConnector (Maybe LogDelivery)

Details about log delivery.

createConnector_workerConfiguration :: Lens' CreateConnector (Maybe WorkerConfiguration)

Specifies which worker configuration to use with the connector.

createConnector_connectorDescription :: Lens' CreateConnector (Maybe Text)

A summary description of the connector.

createConnector_capacity :: Lens' CreateConnector Capacity

Information about the capacity allocated to the connector. Exactly one of the two properties must be specified.

createConnector_connectorConfiguration :: Lens' CreateConnector (HashMap Text Text)

A map of keys to values that represent the configuration for the connector.

createConnector_connectorName :: Lens' CreateConnector Text

The name of the connector.

createConnector_kafkaCluster :: Lens' CreateConnector KafkaCluster

Specifies which Apache Kafka cluster to connect to.

createConnector_kafkaClusterClientAuthentication :: Lens' CreateConnector KafkaClusterClientAuthentication

Details of the client authentication used by the Apache Kafka cluster.

createConnector_kafkaClusterEncryptionInTransit :: Lens' CreateConnector KafkaClusterEncryptionInTransit

Details of encryption in transit to the Apache Kafka cluster.

createConnector_kafkaConnectVersion :: Lens' CreateConnector Text

The version of Kafka Connect. It has to be compatible with both the Apache Kafka cluster's version and the plugins.

createConnector_plugins :: Lens' CreateConnector [Plugin]

Specifies which plugins to use for the connector.

createConnector_serviceExecutionRoleArn :: Lens' CreateConnector Text

The Amazon Resource Name (ARN) of the IAM role used by the connector to access the Amazon Web Services resources that it needs. The types of resources depends on the logic of the connector. For example, a connector that has Amazon S3 as a destination must have permissions that allow it to write to the S3 destination bucket.

createConnectorResponse_connectorArn :: Lens' CreateConnectorResponse (Maybe Text)

The Amazon Resource Name (ARN) that Amazon assigned to the connector.

createConnectorResponse_connectorName :: Lens' CreateConnectorResponse (Maybe Text)

The name of the connector.

createConnectorResponse_connectorState :: Lens' CreateConnectorResponse (Maybe ConnectorState)

The state of the connector.

createConnectorResponse_httpStatus :: Lens' CreateConnectorResponse Int

The response's http status code.

DescribeWorkerConfiguration

describeWorkerConfiguration_workerConfigurationArn :: Lens' DescribeWorkerConfiguration Text

The Amazon Resource Name (ARN) of the worker configuration that you want to get information about.

describeWorkerConfigurationResponse_creationTime :: Lens' DescribeWorkerConfigurationResponse (Maybe UTCTime)

The time that the worker configuration was created.

describeWorkerConfigurationResponse_latestRevision :: Lens' DescribeWorkerConfigurationResponse (Maybe WorkerConfigurationRevisionDescription)

The latest revision of the custom configuration.

describeWorkerConfigurationResponse_name :: Lens' DescribeWorkerConfigurationResponse (Maybe Text)

The name of the worker configuration.

describeWorkerConfigurationResponse_workerConfigurationArn :: Lens' DescribeWorkerConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the custom configuration.

describeWorkerConfigurationResponse_description :: Lens' DescribeWorkerConfigurationResponse (Maybe Text)

The description of the worker configuration.

describeWorkerConfigurationResponse_httpStatus :: Lens' DescribeWorkerConfigurationResponse Int

The response's http status code.

DescribeConnector

describeConnector_connectorArn :: Lens' DescribeConnector Text

The Amazon Resource Name (ARN) of the connector that you want to describe.

describeConnectorResponse_creationTime :: Lens' DescribeConnectorResponse (Maybe UTCTime)

The time the connector was created.

describeConnectorResponse_kafkaCluster :: Lens' DescribeConnectorResponse (Maybe KafkaClusterDescription)

The Apache Kafka cluster that the connector is connected to.

describeConnectorResponse_kafkaConnectVersion :: Lens' DescribeConnectorResponse (Maybe Text)

The version of Kafka Connect. It has to be compatible with both the Apache Kafka cluster's version and the plugins.

describeConnectorResponse_logDelivery :: Lens' DescribeConnectorResponse (Maybe LogDeliveryDescription)

Details about delivering logs to Amazon CloudWatch Logs.

describeConnectorResponse_currentVersion :: Lens' DescribeConnectorResponse (Maybe Text)

The current version of the connector.

describeConnectorResponse_connectorConfiguration :: Lens' DescribeConnectorResponse (Maybe (HashMap Text Text))

A map of keys to values that represent the configuration for the connector.

describeConnectorResponse_workerConfiguration :: Lens' DescribeConnectorResponse (Maybe WorkerConfigurationDescription)

Specifies which worker configuration was used for the connector.

describeConnectorResponse_connectorArn :: Lens' DescribeConnectorResponse (Maybe Text)

The Amazon Resource Name (ARN) of the connector.

describeConnectorResponse_connectorName :: Lens' DescribeConnectorResponse (Maybe Text)

The name of the connector.

describeConnectorResponse_connectorState :: Lens' DescribeConnectorResponse (Maybe ConnectorState)

The state of the connector.

describeConnectorResponse_capacity :: Lens' DescribeConnectorResponse (Maybe CapacityDescription)

Information about the capacity of the connector, whether it is auto scaled or provisioned.

describeConnectorResponse_plugins :: Lens' DescribeConnectorResponse (Maybe [PluginDescription])

Specifies which plugins were used for this connector.

describeConnectorResponse_connectorDescription :: Lens' DescribeConnectorResponse (Maybe Text)

A summary description of the connector.

describeConnectorResponse_kafkaClusterClientAuthentication :: Lens' DescribeConnectorResponse (Maybe KafkaClusterClientAuthenticationDescription)

The type of client authentication used to connect to the Apache Kafka cluster. The value is NONE when no client authentication is used.

describeConnectorResponse_kafkaClusterEncryptionInTransit :: Lens' DescribeConnectorResponse (Maybe KafkaClusterEncryptionInTransitDescription)

Details of encryption in transit to the Apache Kafka cluster.

describeConnectorResponse_serviceExecutionRoleArn :: Lens' DescribeConnectorResponse (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role used by the connector to access Amazon Web Services resources.

describeConnectorResponse_httpStatus :: Lens' DescribeConnectorResponse Int

The response's http status code.

CreateCustomPlugin

createCustomPlugin_description :: Lens' CreateCustomPlugin (Maybe Text)

A summary description of the custom plugin.

createCustomPlugin_contentType :: Lens' CreateCustomPlugin CustomPluginContentType

The type of the plugin file.

createCustomPlugin_location :: Lens' CreateCustomPlugin CustomPluginLocation

Information about the location of a custom plugin.

createCustomPlugin_name :: Lens' CreateCustomPlugin Text

The name of the custom plugin.

createCustomPluginResponse_name :: Lens' CreateCustomPluginResponse (Maybe Text)

The name of the custom plugin.

createCustomPluginResponse_customPluginArn :: Lens' CreateCustomPluginResponse (Maybe Text)

The Amazon Resource Name (ARN) that Amazon assigned to the custom plugin.

createCustomPluginResponse_customPluginState :: Lens' CreateCustomPluginResponse (Maybe CustomPluginState)

The state of the custom plugin.

createCustomPluginResponse_revision :: Lens' CreateCustomPluginResponse (Maybe Integer)

The revision of the custom plugin.

createCustomPluginResponse_httpStatus :: Lens' CreateCustomPluginResponse Int

The response's http status code.

Types

ApacheKafkaCluster

apacheKafkaCluster_bootstrapServers :: Lens' ApacheKafkaCluster Text

The bootstrap servers of the cluster.

apacheKafkaCluster_vpc :: Lens' ApacheKafkaCluster Vpc

Details of an Amazon VPC which has network connectivity to the Apache Kafka cluster.

ApacheKafkaClusterDescription

apacheKafkaClusterDescription_bootstrapServers :: Lens' ApacheKafkaClusterDescription (Maybe Text)

The bootstrap servers of the cluster.

apacheKafkaClusterDescription_vpc :: Lens' ApacheKafkaClusterDescription (Maybe VpcDescription)

Details of an Amazon VPC which has network connectivity to the Apache Kafka cluster.

AutoScaling

autoScaling_scaleInPolicy :: Lens' AutoScaling (Maybe ScaleInPolicy)

The sacle-in policy for the connector.

autoScaling_scaleOutPolicy :: Lens' AutoScaling (Maybe ScaleOutPolicy)

The sacle-out policy for the connector.

autoScaling_maxWorkerCount :: Lens' AutoScaling Natural

The maximum number of workers allocated to the connector.

autoScaling_mcuCount :: Lens' AutoScaling Natural

The number of microcontroller units (MCUs) allocated to each connector worker. The valid values are 1,2,4,8.

autoScaling_minWorkerCount :: Lens' AutoScaling Natural

The minimum number of workers allocated to the connector.

AutoScalingDescription

autoScalingDescription_scaleInPolicy :: Lens' AutoScalingDescription (Maybe ScaleInPolicyDescription)

The sacle-in policy for the connector.

autoScalingDescription_mcuCount :: Lens' AutoScalingDescription (Maybe Int)

The number of microcontroller units (MCUs) allocated to each connector worker. The valid values are 1,2,4,8.

autoScalingDescription_scaleOutPolicy :: Lens' AutoScalingDescription (Maybe ScaleOutPolicyDescription)

The sacle-out policy for the connector.>

autoScalingDescription_maxWorkerCount :: Lens' AutoScalingDescription (Maybe Int)

The maximum number of workers allocated to the connector.

autoScalingDescription_minWorkerCount :: Lens' AutoScalingDescription (Maybe Int)

The minimum number of workers allocated to the connector.

AutoScalingUpdate

autoScalingUpdate_maxWorkerCount :: Lens' AutoScalingUpdate Natural

The target maximum number of workers allocated to the connector.

autoScalingUpdate_mcuCount :: Lens' AutoScalingUpdate Natural

The target number of microcontroller units (MCUs) allocated to each connector worker. The valid values are 1,2,4,8.

autoScalingUpdate_minWorkerCount :: Lens' AutoScalingUpdate Natural

The target minimum number of workers allocated to the connector.

autoScalingUpdate_scaleInPolicy :: Lens' AutoScalingUpdate ScaleInPolicyUpdate

The target sacle-in policy for the connector.

autoScalingUpdate_scaleOutPolicy :: Lens' AutoScalingUpdate ScaleOutPolicyUpdate

The target sacle-out policy for the connector.

Capacity

capacity_autoScaling :: Lens' Capacity (Maybe AutoScaling)

Information about the auto scaling parameters for the connector.

capacity_provisionedCapacity :: Lens' Capacity (Maybe ProvisionedCapacity)

Details about a fixed capacity allocated to a connector.

CapacityDescription

capacityDescription_autoScaling :: Lens' CapacityDescription (Maybe AutoScalingDescription)

Describes the connector's auto scaling capacity.

capacityDescription_provisionedCapacity :: Lens' CapacityDescription (Maybe ProvisionedCapacityDescription)

Describes a connector's provisioned capacity.

CapacityUpdate

capacityUpdate_autoScaling :: Lens' CapacityUpdate (Maybe AutoScalingUpdate)

The target auto scaling setting.

capacityUpdate_provisionedCapacity :: Lens' CapacityUpdate (Maybe ProvisionedCapacityUpdate)

The target settings for provisioned capacity.

CloudWatchLogsLogDelivery

cloudWatchLogsLogDelivery_logGroup :: Lens' CloudWatchLogsLogDelivery (Maybe Text)

The name of the CloudWatch log group that is the destination for log delivery.

cloudWatchLogsLogDelivery_enabled :: Lens' CloudWatchLogsLogDelivery Bool

Whether log delivery to Amazon CloudWatch Logs is enabled.

CloudWatchLogsLogDeliveryDescription

cloudWatchLogsLogDeliveryDescription_enabled :: Lens' CloudWatchLogsLogDeliveryDescription (Maybe Bool)

Whether log delivery to Amazon CloudWatch Logs is enabled.

cloudWatchLogsLogDeliveryDescription_logGroup :: Lens' CloudWatchLogsLogDeliveryDescription (Maybe Text)

The name of the CloudWatch log group that is the destination for log delivery.

ConnectorSummary

connectorSummary_creationTime :: Lens' ConnectorSummary (Maybe UTCTime)

The time that the connector was created.

connectorSummary_kafkaCluster :: Lens' ConnectorSummary (Maybe KafkaClusterDescription)

The details of the Apache Kafka cluster to which the connector is connected.

connectorSummary_kafkaConnectVersion :: Lens' ConnectorSummary (Maybe Text)

The version of Kafka Connect. It has to be compatible with both the Apache Kafka cluster's version and the plugins.

connectorSummary_logDelivery :: Lens' ConnectorSummary (Maybe LogDeliveryDescription)

The settings for delivering connector logs to Amazon CloudWatch Logs.

connectorSummary_currentVersion :: Lens' ConnectorSummary (Maybe Text)

The current version of the connector.

connectorSummary_workerConfiguration :: Lens' ConnectorSummary (Maybe WorkerConfigurationDescription)

The worker configurations that are in use with the connector.

connectorSummary_connectorArn :: Lens' ConnectorSummary (Maybe Text)

The Amazon Resource Name (ARN) of the connector.

connectorSummary_connectorName :: Lens' ConnectorSummary (Maybe Text)

The name of the connector.

connectorSummary_connectorState :: Lens' ConnectorSummary (Maybe ConnectorState)

The state of the connector.

connectorSummary_capacity :: Lens' ConnectorSummary (Maybe CapacityDescription)

The connector's compute capacity settings.

connectorSummary_plugins :: Lens' ConnectorSummary (Maybe [PluginDescription])

Specifies which plugins were used for this connector.

connectorSummary_connectorDescription :: Lens' ConnectorSummary (Maybe Text)

The description of the connector.

connectorSummary_kafkaClusterClientAuthentication :: Lens' ConnectorSummary (Maybe KafkaClusterClientAuthenticationDescription)

The type of client authentication used to connect to the Apache Kafka cluster. The value is NONE when no client authentication is used.

connectorSummary_kafkaClusterEncryptionInTransit :: Lens' ConnectorSummary (Maybe KafkaClusterEncryptionInTransitDescription)

Details of encryption in transit to the Apache Kafka cluster.

connectorSummary_serviceExecutionRoleArn :: Lens' ConnectorSummary (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role used by the connector to access Amazon Web Services resources.

CustomPlugin

customPlugin_customPluginArn :: Lens' CustomPlugin Text

The Amazon Resource Name (ARN) of the custom plugin.

customPlugin_revision :: Lens' CustomPlugin Natural

The revision of the custom plugin.

CustomPluginDescription

customPluginDescription_customPluginArn :: Lens' CustomPluginDescription (Maybe Text)

The Amazon Resource Name (ARN) of the custom plugin.

customPluginDescription_revision :: Lens' CustomPluginDescription (Maybe Integer)

The revision of the custom plugin.

CustomPluginFileDescription

customPluginFileDescription_fileMd5 :: Lens' CustomPluginFileDescription (Maybe Text)

The hex-encoded MD5 checksum of the custom plugin file. You can use it to validate the file.

customPluginFileDescription_fileSize :: Lens' CustomPluginFileDescription (Maybe Integer)

The size in bytes of the custom plugin file. You can use it to validate the file.

CustomPluginLocation

customPluginLocation_s3Location :: Lens' CustomPluginLocation S3Location

The S3 bucket Amazon Resource Name (ARN), file key, and object version of the plugin file stored in Amazon S3.

CustomPluginLocationDescription

customPluginLocationDescription_s3Location :: Lens' CustomPluginLocationDescription (Maybe S3LocationDescription)

The S3 bucket Amazon Resource Name (ARN), file key, and object version of the plugin file stored in Amazon S3.

CustomPluginRevisionSummary

customPluginRevisionSummary_creationTime :: Lens' CustomPluginRevisionSummary (Maybe UTCTime)

The time that the custom plugin was created.

customPluginRevisionSummary_location :: Lens' CustomPluginRevisionSummary (Maybe CustomPluginLocationDescription)

Information about the location of the custom plugin.

customPluginRevisionSummary_fileDescription :: Lens' CustomPluginRevisionSummary (Maybe CustomPluginFileDescription)

Details about the custom plugin file.

customPluginRevisionSummary_revision :: Lens' CustomPluginRevisionSummary (Maybe Integer)

The revision of the custom plugin.

customPluginRevisionSummary_description :: Lens' CustomPluginRevisionSummary (Maybe Text)

The description of the custom plugin.

customPluginRevisionSummary_contentType :: Lens' CustomPluginRevisionSummary (Maybe CustomPluginContentType)

The format of the plugin file.

CustomPluginSummary

customPluginSummary_creationTime :: Lens' CustomPluginSummary (Maybe UTCTime)

The time that the custom plugin was created.

customPluginSummary_latestRevision :: Lens' CustomPluginSummary (Maybe CustomPluginRevisionSummary)

The latest revision of the custom plugin.

customPluginSummary_name :: Lens' CustomPluginSummary (Maybe Text)

The name of the custom plugin.

customPluginSummary_customPluginArn :: Lens' CustomPluginSummary (Maybe Text)

The Amazon Resource Name (ARN) of the custom plugin.

customPluginSummary_customPluginState :: Lens' CustomPluginSummary (Maybe CustomPluginState)

The state of the custom plugin.

customPluginSummary_description :: Lens' CustomPluginSummary (Maybe Text)

A description of the custom plugin.

FirehoseLogDelivery

firehoseLogDelivery_deliveryStream :: Lens' FirehoseLogDelivery (Maybe Text)

The name of the Kinesis Data Firehose delivery stream that is the destination for log delivery.

firehoseLogDelivery_enabled :: Lens' FirehoseLogDelivery Bool

Specifies whether connector logs get delivered to Amazon Kinesis Data Firehose.

FirehoseLogDeliveryDescription

firehoseLogDeliveryDescription_enabled :: Lens' FirehoseLogDeliveryDescription (Maybe Bool)

Specifies whether connector logs get delivered to Amazon Kinesis Data Firehose.

firehoseLogDeliveryDescription_deliveryStream :: Lens' FirehoseLogDeliveryDescription (Maybe Text)

The name of the Kinesis Data Firehose delivery stream that is the destination for log delivery.

KafkaCluster

kafkaCluster_apacheKafkaCluster :: Lens' KafkaCluster ApacheKafkaCluster

The Apache Kafka cluster to which the connector is connected.

KafkaClusterClientAuthentication

kafkaClusterClientAuthentication_authenticationType :: Lens' KafkaClusterClientAuthentication KafkaClusterClientAuthenticationType

The type of client authentication used to connect to the Apache Kafka cluster. Value NONE means that no client authentication is used.

KafkaClusterClientAuthenticationDescription

kafkaClusterClientAuthenticationDescription_authenticationType :: Lens' KafkaClusterClientAuthenticationDescription (Maybe KafkaClusterClientAuthenticationType)

The type of client authentication used to connect to the Apache Kafka cluster. Value NONE means that no client authentication is used.

KafkaClusterDescription

kafkaClusterDescription_apacheKafkaCluster :: Lens' KafkaClusterDescription (Maybe ApacheKafkaClusterDescription)

The Apache Kafka cluster to which the connector is connected.

KafkaClusterEncryptionInTransit

kafkaClusterEncryptionInTransit_encryptionType :: Lens' KafkaClusterEncryptionInTransit KafkaClusterEncryptionInTransitType

The type of encryption in transit to the Apache Kafka cluster.

KafkaClusterEncryptionInTransitDescription

kafkaClusterEncryptionInTransitDescription_encryptionType :: Lens' KafkaClusterEncryptionInTransitDescription (Maybe KafkaClusterEncryptionInTransitType)

The type of encryption in transit to the Apache Kafka cluster.

LogDelivery

logDelivery_workerLogDelivery :: Lens' LogDelivery WorkerLogDelivery

The workers can send worker logs to different destination types. This configuration specifies the details of these destinations.

LogDeliveryDescription

logDeliveryDescription_workerLogDelivery :: Lens' LogDeliveryDescription (Maybe WorkerLogDeliveryDescription)

The workers can send worker logs to different destination types. This configuration specifies the details of these destinations.

Plugin

plugin_customPlugin :: Lens' Plugin CustomPlugin

Details about a custom plugin.

PluginDescription

pluginDescription_customPlugin :: Lens' PluginDescription (Maybe CustomPluginDescription)

Details about a custom plugin.

ProvisionedCapacity

provisionedCapacity_mcuCount :: Lens' ProvisionedCapacity Natural

The number of microcontroller units (MCUs) allocated to each connector worker. The valid values are 1,2,4,8.

provisionedCapacity_workerCount :: Lens' ProvisionedCapacity Natural

The number of workers that are allocated to the connector.

ProvisionedCapacityDescription

provisionedCapacityDescription_workerCount :: Lens' ProvisionedCapacityDescription (Maybe Int)

The number of workers that are allocated to the connector.

provisionedCapacityDescription_mcuCount :: Lens' ProvisionedCapacityDescription (Maybe Int)

The number of microcontroller units (MCUs) allocated to each connector worker. The valid values are 1,2,4,8.

ProvisionedCapacityUpdate

provisionedCapacityUpdate_mcuCount :: Lens' ProvisionedCapacityUpdate Natural

The number of microcontroller units (MCUs) allocated to each connector worker. The valid values are 1,2,4,8.

provisionedCapacityUpdate_workerCount :: Lens' ProvisionedCapacityUpdate Natural

The number of workers that are allocated to the connector.

S3Location

s3Location_objectVersion :: Lens' S3Location (Maybe Text)

The version of an object in an S3 bucket.

s3Location_bucketArn :: Lens' S3Location Text

The Amazon Resource Name (ARN) of an S3 bucket.

s3Location_fileKey :: Lens' S3Location Text

The file key for an object in an S3 bucket.

S3LocationDescription

s3LocationDescription_objectVersion :: Lens' S3LocationDescription (Maybe Text)

The version of an object in an S3 bucket.

s3LocationDescription_bucketArn :: Lens' S3LocationDescription (Maybe Text)

The Amazon Resource Name (ARN) of an S3 bucket.

s3LocationDescription_fileKey :: Lens' S3LocationDescription (Maybe Text)

The file key for an object in an S3 bucket.

S3LogDelivery

s3LogDelivery_prefix :: Lens' S3LogDelivery (Maybe Text)

The S3 prefix that is the destination for log delivery.

s3LogDelivery_bucket :: Lens' S3LogDelivery (Maybe Text)

The name of the S3 bucket that is the destination for log delivery.

s3LogDelivery_enabled :: Lens' S3LogDelivery Bool

Specifies whether connector logs get sent to the specified Amazon S3 destination.

S3LogDeliveryDescription

s3LogDeliveryDescription_enabled :: Lens' S3LogDeliveryDescription (Maybe Bool)

Specifies whether connector logs get sent to the specified Amazon S3 destination.

s3LogDeliveryDescription_prefix :: Lens' S3LogDeliveryDescription (Maybe Text)

The S3 prefix that is the destination for log delivery.

s3LogDeliveryDescription_bucket :: Lens' S3LogDeliveryDescription (Maybe Text)

The name of the S3 bucket that is the destination for log delivery.

ScaleInPolicy

scaleInPolicy_cpuUtilizationPercentage :: Lens' ScaleInPolicy Natural

Specifies the CPU utilization percentage threshold at which you want connector scale in to be triggered.

ScaleInPolicyDescription

scaleInPolicyDescription_cpuUtilizationPercentage :: Lens' ScaleInPolicyDescription (Maybe Int)

Specifies the CPU utilization percentage threshold at which you want connector scale in to be triggered.

ScaleInPolicyUpdate

scaleInPolicyUpdate_cpuUtilizationPercentage :: Lens' ScaleInPolicyUpdate Natural

The target CPU utilization percentage threshold at which you want connector scale in to be triggered.

ScaleOutPolicy

scaleOutPolicy_cpuUtilizationPercentage :: Lens' ScaleOutPolicy Natural

The CPU utilization percentage threshold at which you want connector scale out to be triggered.

ScaleOutPolicyDescription

scaleOutPolicyDescription_cpuUtilizationPercentage :: Lens' ScaleOutPolicyDescription (Maybe Int)

The CPU utilization percentage threshold at which you want connector scale out to be triggered.

ScaleOutPolicyUpdate

scaleOutPolicyUpdate_cpuUtilizationPercentage :: Lens' ScaleOutPolicyUpdate Natural

The target CPU utilization percentage threshold at which you want connector scale out to be triggered.

Vpc

vpc_securityGroups :: Lens' Vpc (Maybe [Text])

The security groups for the connector.

vpc_subnets :: Lens' Vpc [Text]

The subnets for the connector.

VpcDescription

vpcDescription_securityGroups :: Lens' VpcDescription (Maybe [Text])

The security groups for the connector.

vpcDescription_subnets :: Lens' VpcDescription (Maybe [Text])

The subnets for the connector.

WorkerConfiguration

workerConfiguration_revision :: Lens' WorkerConfiguration Natural

The revision of the worker configuration.

workerConfiguration_workerConfigurationArn :: Lens' WorkerConfiguration Text

The Amazon Resource Name (ARN) of the worker configuration.

WorkerConfigurationDescription

workerConfigurationDescription_workerConfigurationArn :: Lens' WorkerConfigurationDescription (Maybe Text)

The Amazon Resource Name (ARN) of the worker configuration.

workerConfigurationDescription_revision :: Lens' WorkerConfigurationDescription (Maybe Integer)

The revision of the worker configuration.

WorkerConfigurationRevisionDescription

workerConfigurationRevisionDescription_creationTime :: Lens' WorkerConfigurationRevisionDescription (Maybe UTCTime)

The time that the worker configuration was created.

workerConfigurationRevisionDescription_revision :: Lens' WorkerConfigurationRevisionDescription (Maybe Integer)

The description of a revision of the worker configuration.

workerConfigurationRevisionDescription_propertiesFileContent :: Lens' WorkerConfigurationRevisionDescription (Maybe Text)

Base64 encoded contents of the connect-distributed.properties file.

workerConfigurationRevisionDescription_description :: Lens' WorkerConfigurationRevisionDescription (Maybe Text)

The description of the worker configuration revision.

WorkerConfigurationRevisionSummary

workerConfigurationRevisionSummary_creationTime :: Lens' WorkerConfigurationRevisionSummary (Maybe UTCTime)

The time that a worker configuration revision was created.

workerConfigurationRevisionSummary_revision :: Lens' WorkerConfigurationRevisionSummary (Maybe Integer)

The revision of a worker configuration.

workerConfigurationRevisionSummary_description :: Lens' WorkerConfigurationRevisionSummary (Maybe Text)

The description of a worker configuration revision.

WorkerConfigurationSummary

workerConfigurationSummary_creationTime :: Lens' WorkerConfigurationSummary (Maybe UTCTime)

The time that a worker configuration was created.

workerConfigurationSummary_latestRevision :: Lens' WorkerConfigurationSummary (Maybe WorkerConfigurationRevisionSummary)

The latest revision of a worker configuration.

workerConfigurationSummary_name :: Lens' WorkerConfigurationSummary (Maybe Text)

The name of the worker configuration.

workerConfigurationSummary_workerConfigurationArn :: Lens' WorkerConfigurationSummary (Maybe Text)

The Amazon Resource Name (ARN) of the worker configuration.

workerConfigurationSummary_description :: Lens' WorkerConfigurationSummary (Maybe Text)

The description of a worker configuration.

WorkerLogDelivery

workerLogDelivery_cloudWatchLogs :: Lens' WorkerLogDelivery (Maybe CloudWatchLogsLogDelivery)

Details about delivering logs to Amazon CloudWatch Logs.

workerLogDelivery_firehose :: Lens' WorkerLogDelivery (Maybe FirehoseLogDelivery)

Details about delivering logs to Amazon Kinesis Data Firehose.

workerLogDelivery_s3 :: Lens' WorkerLogDelivery (Maybe S3LogDelivery)

Details about delivering logs to Amazon S3.

WorkerLogDeliveryDescription

workerLogDeliveryDescription_cloudWatchLogs :: Lens' WorkerLogDeliveryDescription (Maybe CloudWatchLogsLogDeliveryDescription)

Details about delivering logs to Amazon CloudWatch Logs.

workerLogDeliveryDescription_firehose :: Lens' WorkerLogDeliveryDescription (Maybe FirehoseLogDeliveryDescription)

Details about delivering logs to Amazon Kinesis Data Firehose.

workerLogDeliveryDescription_s3 :: Lens' WorkerLogDeliveryDescription (Maybe S3LogDeliveryDescription)

Details about delivering logs to Amazon S3.