Network.AWS.KafkaConnect.ListWorkerConfigurations

Description

Returns a list of all of the worker configurations in this account and Region.

This operation returns paginated results.

Creating a Request

data ListWorkerConfigurations

See: newListWorkerConfigurations smart constructor.

Constructors

ListWorkerConfigurations'

Fields

nextToken :: Maybe Text

If the response of a ListWorkerConfigurations operation is truncated, it will include a NextToken. Send this NextToken in a subsequent request to continue listing from where the previous operation left off.

maxResults :: Maybe Natural

The maximum number of worker configurations to list in one response.

newListWorkerConfigurations :: ListWorkerConfigurations

Create a value of ListWorkerConfigurations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListWorkerConfigurations', listWorkerConfigurations_nextToken - If the response of a ListWorkerConfigurations operation is truncated, it will include a NextToken. Send this NextToken in a subsequent request to continue listing from where the previous operation left off.

$sel:maxResults:ListWorkerConfigurations', listWorkerConfigurations_maxResults - The maximum number of worker configurations to list in one response.

Request Lenses

listWorkerConfigurations_nextToken :: Lens' ListWorkerConfigurations (Maybe Text)

If the response of a ListWorkerConfigurations operation is truncated, it will include a NextToken. Send this NextToken in a subsequent request to continue listing from where the previous operation left off.

listWorkerConfigurations_maxResults :: Lens' ListWorkerConfigurations (Maybe Natural)

The maximum number of worker configurations to list in one response.

Destructuring the Response

data ListWorkerConfigurationsResponse

See: newListWorkerConfigurationsResponse smart constructor.

Constructors

ListWorkerConfigurationsResponse'

Fields

workerConfigurations :: Maybe [WorkerConfigurationSummary]

An array of worker configuration descriptions.

nextToken :: Maybe Text

If the response of a ListWorkerConfigurations operation is truncated, it will include a NextToken. Send this NextToken in a subsequent request to continue listing from where the previous operation left off.

httpStatus :: Int

The response's http status code.

newListWorkerConfigurationsResponse

Arguments

:: Int
$sel:httpStatus:ListWorkerConfigurationsResponse'

-> ListWorkerConfigurationsResponse

Create a value of ListWorkerConfigurationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:workerConfigurations:ListWorkerConfigurationsResponse', listWorkerConfigurationsResponse_workerConfigurations - An array of worker configuration descriptions.

$sel:nextToken:ListWorkerConfigurations', listWorkerConfigurationsResponse_nextToken - If the response of a ListWorkerConfigurations operation is truncated, it will include a NextToken. Send this NextToken in a subsequent request to continue listing from where the previous operation left off.

$sel:httpStatus:ListWorkerConfigurationsResponse', listWorkerConfigurationsResponse_httpStatus - The response's http status code.

Response Lenses

listWorkerConfigurationsResponse_workerConfigurations :: Lens' ListWorkerConfigurationsResponse (Maybe [WorkerConfigurationSummary])

An array of worker configuration descriptions.

listWorkerConfigurationsResponse_nextToken :: Lens' ListWorkerConfigurationsResponse (Maybe Text)

If the response of a ListWorkerConfigurations operation is truncated, it will include a NextToken. Send this NextToken in a subsequent request to continue listing from where the previous operation left off.

listWorkerConfigurationsResponse_httpStatus :: Lens' ListWorkerConfigurationsResponse Int

The response's http status code.