Network.AWS.KafkaConnect.Types.Capacity

Description

Documentation

data Capacity

Information about the capacity of the connector, whether it is auto scaled or provisioned.

See: newCapacity smart constructor.

Constructors

Capacity'

Fields

autoScaling :: Maybe AutoScaling

Information about the auto scaling parameters for the connector.

provisionedCapacity :: Maybe ProvisionedCapacity

Details about a fixed capacity allocated to a connector.

newCapacity :: Capacity

Create a value of Capacity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoScaling:Capacity', capacity_autoScaling - Information about the auto scaling parameters for the connector.

$sel:provisionedCapacity:Capacity', capacity_provisionedCapacity - Details about a fixed capacity allocated to a connector.

capacity_autoScaling :: Lens' Capacity (Maybe AutoScaling)

Information about the auto scaling parameters for the connector.

capacity_provisionedCapacity :: Lens' Capacity (Maybe ProvisionedCapacity)

Details about a fixed capacity allocated to a connector.