Network.AWS.KafkaConnect.Types.ConnectorSummary

Description

Documentation

data ConnectorSummary

Summary of a connector.

See: newConnectorSummary smart constructor.

Constructors

ConnectorSummary'

Fields

creationTime :: Maybe POSIX

The time that the connector was created.

kafkaCluster :: Maybe KafkaClusterDescription

The details of the Apache Kafka cluster to which the connector is connected.

kafkaConnectVersion :: Maybe Text

The version of Kafka Connect. It has to be compatible with both the Apache Kafka cluster's version and the plugins.

logDelivery :: Maybe LogDeliveryDescription

The settings for delivering connector logs to Amazon CloudWatch Logs.

currentVersion :: Maybe Text

The current version of the connector.

workerConfiguration :: Maybe WorkerConfigurationDescription

The worker configurations that are in use with the connector.

connectorArn :: Maybe Text

The Amazon Resource Name (ARN) of the connector.

connectorName :: Maybe Text

The name of the connector.

connectorState :: Maybe ConnectorState

The state of the connector.

capacity :: Maybe CapacityDescription

The connector's compute capacity settings.

plugins :: Maybe [PluginDescription]

Specifies which plugins were used for this connector.

connectorDescription :: Maybe Text

The description of the connector.

kafkaClusterClientAuthentication :: Maybe KafkaClusterClientAuthenticationDescription

The type of client authentication used to connect to the Apache Kafka cluster. The value is NONE when no client authentication is used.

kafkaClusterEncryptionInTransit :: Maybe KafkaClusterEncryptionInTransitDescription

Details of encryption in transit to the Apache Kafka cluster.

serviceExecutionRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role used by the connector to access Amazon Web Services resources.

newConnectorSummary :: ConnectorSummary

Create a value of ConnectorSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ConnectorSummary', connectorSummary_creationTime - The time that the connector was created.

$sel:kafkaCluster:ConnectorSummary', connectorSummary_kafkaCluster - The details of the Apache Kafka cluster to which the connector is connected.

$sel:kafkaConnectVersion:ConnectorSummary', connectorSummary_kafkaConnectVersion - The version of Kafka Connect. It has to be compatible with both the Apache Kafka cluster's version and the plugins.

$sel:logDelivery:ConnectorSummary', connectorSummary_logDelivery - The settings for delivering connector logs to Amazon CloudWatch Logs.

$sel:currentVersion:ConnectorSummary', connectorSummary_currentVersion - The current version of the connector.

$sel:workerConfiguration:ConnectorSummary', connectorSummary_workerConfiguration - The worker configurations that are in use with the connector.

$sel:connectorArn:ConnectorSummary', connectorSummary_connectorArn - The Amazon Resource Name (ARN) of the connector.

$sel:connectorName:ConnectorSummary', connectorSummary_connectorName - The name of the connector.

$sel:connectorState:ConnectorSummary', connectorSummary_connectorState - The state of the connector.

$sel:capacity:ConnectorSummary', connectorSummary_capacity - The connector's compute capacity settings.

$sel:plugins:ConnectorSummary', connectorSummary_plugins - Specifies which plugins were used for this connector.

$sel:connectorDescription:ConnectorSummary', connectorSummary_connectorDescription - The description of the connector.

$sel:kafkaClusterClientAuthentication:ConnectorSummary', connectorSummary_kafkaClusterClientAuthentication - The type of client authentication used to connect to the Apache Kafka cluster. The value is NONE when no client authentication is used.

$sel:kafkaClusterEncryptionInTransit:ConnectorSummary', connectorSummary_kafkaClusterEncryptionInTransit - Details of encryption in transit to the Apache Kafka cluster.

$sel:serviceExecutionRoleArn:ConnectorSummary', connectorSummary_serviceExecutionRoleArn - The Amazon Resource Name (ARN) of the IAM role used by the connector to access Amazon Web Services resources.

connectorSummary_creationTime :: Lens' ConnectorSummary (Maybe UTCTime)

The time that the connector was created.

connectorSummary_kafkaCluster :: Lens' ConnectorSummary (Maybe KafkaClusterDescription)

The details of the Apache Kafka cluster to which the connector is connected.

connectorSummary_kafkaConnectVersion :: Lens' ConnectorSummary (Maybe Text)

The version of Kafka Connect. It has to be compatible with both the Apache Kafka cluster's version and the plugins.

connectorSummary_logDelivery :: Lens' ConnectorSummary (Maybe LogDeliveryDescription)

The settings for delivering connector logs to Amazon CloudWatch Logs.

connectorSummary_currentVersion :: Lens' ConnectorSummary (Maybe Text)

The current version of the connector.

connectorSummary_workerConfiguration :: Lens' ConnectorSummary (Maybe WorkerConfigurationDescription)

The worker configurations that are in use with the connector.

connectorSummary_connectorArn :: Lens' ConnectorSummary (Maybe Text)

The Amazon Resource Name (ARN) of the connector.

connectorSummary_connectorName :: Lens' ConnectorSummary (Maybe Text)

The name of the connector.

connectorSummary_connectorState :: Lens' ConnectorSummary (Maybe ConnectorState)

The state of the connector.

connectorSummary_capacity :: Lens' ConnectorSummary (Maybe CapacityDescription)

The connector's compute capacity settings.

connectorSummary_plugins :: Lens' ConnectorSummary (Maybe [PluginDescription])

Specifies which plugins were used for this connector.

connectorSummary_connectorDescription :: Lens' ConnectorSummary (Maybe Text)

The description of the connector.

connectorSummary_kafkaClusterClientAuthentication :: Lens' ConnectorSummary (Maybe KafkaClusterClientAuthenticationDescription)

The type of client authentication used to connect to the Apache Kafka cluster. The value is NONE when no client authentication is used.

connectorSummary_kafkaClusterEncryptionInTransit :: Lens' ConnectorSummary (Maybe KafkaClusterEncryptionInTransitDescription)

Details of encryption in transit to the Apache Kafka cluster.

connectorSummary_serviceExecutionRoleArn :: Lens' ConnectorSummary (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role used by the connector to access Amazon Web Services resources.