Network.AWS.KafkaConnect.Types.Plugin

Description

Documentation

data Plugin

A plugin is an AWS resource that contains the code that defines your connector logic.

See: newPlugin smart constructor.

Constructors

Plugin'

Fields

customPlugin :: CustomPlugin

Details about a custom plugin.

newPlugin

Arguments

:: CustomPlugin
$sel:customPlugin:Plugin'

-> Plugin

Create a value of Plugin with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customPlugin:Plugin', plugin_customPlugin - Details about a custom plugin.

plugin_customPlugin :: Lens' Plugin CustomPlugin

Details about a custom plugin.