Network.AWS.KafkaConnect.Types.WorkerConfigurationRevisionSummary

Description

Documentation

data WorkerConfigurationRevisionSummary

The summary of a worker configuration revision.

See: newWorkerConfigurationRevisionSummary smart constructor.

Constructors

WorkerConfigurationRevisionSummary'

Fields

creationTime :: Maybe POSIX

The time that a worker configuration revision was created.

revision :: Maybe Integer

The revision of a worker configuration.

description :: Maybe Text

The description of a worker configuration revision.

newWorkerConfigurationRevisionSummary :: WorkerConfigurationRevisionSummary

Create a value of WorkerConfigurationRevisionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:WorkerConfigurationRevisionSummary', workerConfigurationRevisionSummary_creationTime - The time that a worker configuration revision was created.

$sel:revision:WorkerConfigurationRevisionSummary', workerConfigurationRevisionSummary_revision - The revision of a worker configuration.

$sel:description:WorkerConfigurationRevisionSummary', workerConfigurationRevisionSummary_description - The description of a worker configuration revision.

workerConfigurationRevisionSummary_creationTime :: Lens' WorkerConfigurationRevisionSummary (Maybe UTCTime)

The time that a worker configuration revision was created.

workerConfigurationRevisionSummary_revision :: Lens' WorkerConfigurationRevisionSummary (Maybe Integer)

The revision of a worker configuration.

workerConfigurationRevisionSummary_description :: Lens' WorkerConfigurationRevisionSummary (Maybe Text)

The description of a worker configuration revision.