Network.AWS.Kinesis.AddTagsToStream

Description

Adds or updates tags for the specified Kinesis data stream. Each time you invoke this operation, you can specify up to 10 tags. If you want to add more than 10 tags to your stream, you can invoke this operation multiple times. In total, each stream can have up to 50 tags.

If tags have already been assigned to the stream, AddTagsToStream overwrites any existing tags that correspond to the specified tag keys.

AddTagsToStream has a limit of five transactions per second per account.

Creating a Request

data AddTagsToStream

Represents the input for AddTagsToStream.

See: newAddTagsToStream smart constructor.

Constructors

AddTagsToStream'

Fields

streamName :: Text

The name of the stream.

tags :: HashMap Text Text

A set of up to 10 key-value pairs to use to create the tags.

newAddTagsToStream

Arguments

:: Text
$sel:streamName:AddTagsToStream'

-> AddTagsToStream

Create a value of AddTagsToStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamName:AddTagsToStream', addTagsToStream_streamName - The name of the stream.

$sel:tags:AddTagsToStream', addTagsToStream_tags - A set of up to 10 key-value pairs to use to create the tags.

Request Lenses

addTagsToStream_streamName :: Lens' AddTagsToStream Text

The name of the stream.

addTagsToStream_tags :: Lens' AddTagsToStream (HashMap Text Text)

A set of up to 10 key-value pairs to use to create the tags.

Destructuring the Response

data AddTagsToStreamResponse

See: newAddTagsToStreamResponse smart constructor.

Constructors

AddTagsToStreamResponse'

newAddTagsToStreamResponse :: AddTagsToStreamResponse

Create a value of AddTagsToStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.