Network.AWS.Kinesis.DecreaseStreamRetentionPeriod

Description

Decreases the Kinesis data stream's retention period, which is the length of time data records are accessible after they are added to the stream. The minimum value of a stream's retention period is 24 hours.

This operation may result in lost data. For example, if the stream's retention period is 48 hours and is decreased to 24 hours, any data already in the stream that is older than 24 hours is inaccessible.

Creating a Request

data DecreaseStreamRetentionPeriod

Represents the input for DecreaseStreamRetentionPeriod.

See: newDecreaseStreamRetentionPeriod smart constructor.

Constructors

DecreaseStreamRetentionPeriod'

Fields

streamName :: Text

The name of the stream to modify.

retentionPeriodHours :: Int

The new retention period of the stream, in hours. Must be less than the current retention period.

newDecreaseStreamRetentionPeriod

Arguments

:: Text
$sel:streamName:DecreaseStreamRetentionPeriod'

-> Int
$sel:retentionPeriodHours:DecreaseStreamRetentionPeriod'

-> DecreaseStreamRetentionPeriod

Create a value of DecreaseStreamRetentionPeriod with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamName:DecreaseStreamRetentionPeriod', decreaseStreamRetentionPeriod_streamName - The name of the stream to modify.

$sel:retentionPeriodHours:DecreaseStreamRetentionPeriod', decreaseStreamRetentionPeriod_retentionPeriodHours - The new retention period of the stream, in hours. Must be less than the current retention period.

Request Lenses

decreaseStreamRetentionPeriod_streamName :: Lens' DecreaseStreamRetentionPeriod Text

The name of the stream to modify.

decreaseStreamRetentionPeriod_retentionPeriodHours :: Lens' DecreaseStreamRetentionPeriod Int

The new retention period of the stream, in hours. Must be less than the current retention period.

Destructuring the Response

data DecreaseStreamRetentionPeriodResponse

See: newDecreaseStreamRetentionPeriodResponse smart constructor.

Constructors

DecreaseStreamRetentionPeriodResponse'

newDecreaseStreamRetentionPeriodResponse :: DecreaseStreamRetentionPeriodResponse

Create a value of DecreaseStreamRetentionPeriodResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.