Network.AWS.Kinesis.DescribeStream

Description

Describes the specified Kinesis data stream.

The information returned includes the stream name, Amazon Resource Name (ARN), creation time, enhanced metric configuration, and shard map. The shard map is an array of shard objects. For each shard object, there is the hash key and sequence number ranges that the shard spans, and the IDs of any earlier shards that played in a role in creating the shard. Every record ingested in the stream is identified by a sequence number, which is assigned when the record is put into the stream.

You can limit the number of shards returned by each call. For more information, see Retrieving Shards from a Stream in the Amazon Kinesis Data Streams Developer Guide.

There are no guarantees about the chronological order shards returned. To process shards in chronological order, use the ID of the parent shard to track the lineage to the oldest shard.

This operation has a limit of 10 transactions per second per account.

This operation returns paginated results.

Creating a Request

data DescribeStream

Represents the input for DescribeStream.

See: newDescribeStream smart constructor.

Constructors

DescribeStream'

Fields

exclusiveStartShardId :: Maybe Text

The shard ID of the shard to start with.

limit :: Maybe Natural

The maximum number of shards to return in a single call. The default value is 100. If you specify a value greater than 100, at most 100 shards are returned.

streamName :: Text

The name of the stream to describe.

newDescribeStream

Arguments

:: Text
$sel:streamName:DescribeStream'

-> DescribeStream

Create a value of DescribeStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:exclusiveStartShardId:DescribeStream', describeStream_exclusiveStartShardId - The shard ID of the shard to start with.

$sel:limit:DescribeStream', describeStream_limit - The maximum number of shards to return in a single call. The default value is 100. If you specify a value greater than 100, at most 100 shards are returned.

$sel:streamName:DescribeStream', describeStream_streamName - The name of the stream to describe.

Request Lenses

describeStream_exclusiveStartShardId :: Lens' DescribeStream (Maybe Text)

The shard ID of the shard to start with.

describeStream_limit :: Lens' DescribeStream (Maybe Natural)

The maximum number of shards to return in a single call. The default value is 100. If you specify a value greater than 100, at most 100 shards are returned.

describeStream_streamName :: Lens' DescribeStream Text

The name of the stream to describe.

Destructuring the Response

data DescribeStreamResponse

Represents the output for DescribeStream.

See: newDescribeStreamResponse smart constructor.

Constructors

DescribeStreamResponse'

Fields

httpStatus :: Int

The response's http status code.

streamDescription :: StreamDescription

The current status of the stream, the stream Amazon Resource Name (ARN), an array of shard objects that comprise the stream, and whether there are more shards available.

newDescribeStreamResponse

Arguments

:: Int
$sel:httpStatus:DescribeStreamResponse'

-> StreamDescription
$sel:streamDescription:DescribeStreamResponse'

-> DescribeStreamResponse

Create a value of DescribeStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeStreamResponse', describeStreamResponse_httpStatus - The response's http status code.

$sel:streamDescription:DescribeStreamResponse', describeStreamResponse_streamDescription - The current status of the stream, the stream Amazon Resource Name (ARN), an array of shard objects that comprise the stream, and whether there are more shards available.

Response Lenses

describeStreamResponse_httpStatus :: Lens' DescribeStreamResponse Int

The response's http status code.

describeStreamResponse_streamDescription :: Lens' DescribeStreamResponse StreamDescription

The current status of the stream, the stream Amazon Resource Name (ARN), an array of shard objects that comprise the stream, and whether there are more shards available.