Network.AWS.Kinesis.GetRecords

Description

Gets data records from a Kinesis data stream's shard.

Specify a shard iterator using the ShardIterator parameter. The shard iterator specifies the position in the shard from which you want to start reading data records sequentially. If there are no records available in the portion of the shard that the iterator points to, GetRecords returns an empty list. It might take multiple calls to get to a portion of the shard that contains records.

You can scale by provisioning multiple shards per stream while considering service limits (for more information, see Amazon Kinesis Data Streams Limits in the Amazon Kinesis Data Streams Developer Guide). Your application should have one thread per shard, each reading continuously from its stream. To read from a stream continually, call GetRecords in a loop. Use GetShardIterator to get the shard iterator to specify in the first GetRecords call. GetRecords returns a new shard iterator in NextShardIterator. Specify the shard iterator returned in NextShardIterator in subsequent calls to GetRecords. If the shard has been closed, the shard iterator can't return more data and GetRecords returns null in NextShardIterator. You can terminate the loop when the shard is closed, or when the shard iterator reaches the record with the sequence number or other attribute that marks it as the last record to process.

Each data record can be up to 1 MiB in size, and each shard can read up to 2 MiB per second. You can ensure that your calls don't exceed the maximum supported size or throughput by using the Limit parameter to specify the maximum number of records that GetRecords can return. Consider your average record size when determining this limit. The maximum number of records that can be returned per call is 10,000.

The size of the data returned by GetRecords varies depending on the utilization of the shard. The maximum size of data that GetRecords can return is 10 MiB. If a call returns this amount of data, subsequent calls made within the next 5 seconds throw ProvisionedThroughputExceededException. If there is insufficient provisioned throughput on the stream, subsequent calls made within the next 1 second throw ProvisionedThroughputExceededException. GetRecords doesn't return any data when it throws an exception. For this reason, we recommend that you wait 1 second between calls to GetRecords. However, it's possible that the application will get exceptions for longer than 1 second.

To detect whether the application is falling behind in processing, you can use the MillisBehindLatest response attribute. You can also monitor the stream using CloudWatch metrics and other mechanisms (see Monitoring in the Amazon Kinesis Data Streams Developer Guide).

Each Amazon Kinesis record includes a value, ApproximateArrivalTimestamp, that is set when a stream successfully receives and stores a record. This is commonly referred to as a server-side time stamp, whereas a client-side time stamp is set when a data producer creates or sends the record to a stream (a data producer is any data source putting data records into a stream, for example with PutRecords). The time stamp has millisecond precision. There are no guarantees about the time stamp accuracy, or that the time stamp is always increasing. For example, records in a shard or across a stream might have time stamps that are out of order.

This operation has a limit of five transactions per second per shard.

Creating a Request

data GetRecords

Represents the input for GetRecords.

See: newGetRecords smart constructor.

Constructors

GetRecords'

Fields

limit :: Maybe Natural

The maximum number of records to return. Specify a value of up to 10,000. If you specify a value that is greater than 10,000, GetRecords throws InvalidArgumentException. The default value is 10,000.

shardIterator :: Text

The position in the shard from which you want to start sequentially reading data records. A shard iterator specifies this position using the sequence number of a data record in the shard.

newGetRecords

Arguments

:: Text
$sel:shardIterator:GetRecords'

-> GetRecords

Create a value of GetRecords with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:limit:GetRecords', getRecords_limit - The maximum number of records to return. Specify a value of up to 10,000. If you specify a value that is greater than 10,000, GetRecords throws InvalidArgumentException. The default value is 10,000.

$sel:shardIterator:GetRecords', getRecords_shardIterator - The position in the shard from which you want to start sequentially reading data records. A shard iterator specifies this position using the sequence number of a data record in the shard.

Request Lenses

getRecords_limit :: Lens' GetRecords (Maybe Natural)

The maximum number of records to return. Specify a value of up to 10,000. If you specify a value that is greater than 10,000, GetRecords throws InvalidArgumentException. The default value is 10,000.

getRecords_shardIterator :: Lens' GetRecords Text

The position in the shard from which you want to start sequentially reading data records. A shard iterator specifies this position using the sequence number of a data record in the shard.

Destructuring the Response

data GetRecordsResponse

Represents the output for GetRecords.

See: newGetRecordsResponse smart constructor.

Constructors

GetRecordsResponse'

Fields

millisBehindLatest :: Maybe Natural

The number of milliseconds the GetRecords response is from the tip of the stream, indicating how far behind current time the consumer is. A value of zero indicates that record processing is caught up, and there are no new records to process at this moment.

nextShardIterator :: Maybe Text

The next position in the shard from which to start sequentially reading data records. If set to null, the shard has been closed and the requested iterator does not return any more data.

childShards :: Maybe [ChildShard]

httpStatus :: Int

The response's http status code.

records :: [Record]

The data records retrieved from the shard.

newGetRecordsResponse

Arguments

:: Int
$sel:httpStatus:GetRecordsResponse'

-> GetRecordsResponse

Create a value of GetRecordsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:millisBehindLatest:GetRecordsResponse', getRecordsResponse_millisBehindLatest - The number of milliseconds the GetRecords response is from the tip of the stream, indicating how far behind current time the consumer is. A value of zero indicates that record processing is caught up, and there are no new records to process at this moment.

$sel:nextShardIterator:GetRecordsResponse', getRecordsResponse_nextShardIterator - The next position in the shard from which to start sequentially reading data records. If set to null, the shard has been closed and the requested iterator does not return any more data.

$sel:childShards:GetRecordsResponse', getRecordsResponse_childShards - Undocumented member.

$sel:httpStatus:GetRecordsResponse', getRecordsResponse_httpStatus - The response's http status code.

$sel:records:GetRecordsResponse', getRecordsResponse_records - The data records retrieved from the shard.

Response Lenses

getRecordsResponse_millisBehindLatest :: Lens' GetRecordsResponse (Maybe Natural)

The number of milliseconds the GetRecords response is from the tip of the stream, indicating how far behind current time the consumer is. A value of zero indicates that record processing is caught up, and there are no new records to process at this moment.

getRecordsResponse_nextShardIterator :: Lens' GetRecordsResponse (Maybe Text)

The next position in the shard from which to start sequentially reading data records. If set to null, the shard has been closed and the requested iterator does not return any more data.

getRecordsResponse_childShards :: Lens' GetRecordsResponse (Maybe [ChildShard])

Undocumented member.

getRecordsResponse_httpStatus :: Lens' GetRecordsResponse Int

The response's http status code.

getRecordsResponse_records :: Lens' GetRecordsResponse [Record]

The data records retrieved from the shard.