Network.AWS.Kinesis.GetShardIterator

Description

Gets an Amazon Kinesis shard iterator. A shard iterator expires 5 minutes after it is returned to the requester.

A shard iterator specifies the shard position from which to start reading data records sequentially. The position is specified using the sequence number of a data record in a shard. A sequence number is the identifier associated with every record ingested in the stream, and is assigned when a record is put into the stream. Each stream has one or more shards.

You must specify the shard iterator type. For example, you can set the ShardIteratorType parameter to read exactly from the position denoted by a specific sequence number by using the AT_SEQUENCE_NUMBER shard iterator type. Alternatively, the parameter can read right after the sequence number by using the AFTER_SEQUENCE_NUMBER shard iterator type, using sequence numbers returned by earlier calls to PutRecord, PutRecords, GetRecords, or DescribeStream. In the request, you can specify the shard iterator type AT_TIMESTAMP to read records from an arbitrary point in time, TRIM_HORIZON to cause ShardIterator to point to the last untrimmed record in the shard in the system (the oldest data record in the shard), or LATEST so that you always read the most recent data in the shard.

When you read repeatedly from a stream, use a GetShardIterator request to get the first shard iterator for use in your first GetRecords request and for subsequent reads use the shard iterator returned by the GetRecords request in NextShardIterator. A new shard iterator is returned by every GetRecords request in NextShardIterator, which you use in the ShardIterator parameter of the next GetRecords request.

If a GetShardIterator request is made too often, you receive a ProvisionedThroughputExceededException. For more information about throughput limits, see GetRecords, and Streams Limits in the Amazon Kinesis Data Streams Developer Guide.

If the shard is closed, GetShardIterator returns a valid iterator for the last sequence number of the shard. A shard can be closed as a result of using SplitShard or MergeShards.

GetShardIterator has a limit of five transactions per second per account per open shard.

Creating a Request

data GetShardIterator

Represents the input for GetShardIterator.

See: newGetShardIterator smart constructor.

Constructors

GetShardIterator'

Fields

startingSequenceNumber :: Maybe Text

The sequence number of the data record in the shard from which to start reading. Used with shard iterator type AT_SEQUENCE_NUMBER and AFTER_SEQUENCE_NUMBER.

timestamp :: Maybe POSIX

The time stamp of the data record from which to start reading. Used with shard iterator type AT_TIMESTAMP. A time stamp is the Unix epoch date with precision in milliseconds. For example, 2016-04-04T19:58:46.480-00:00 or 1459799926.480. If a record with this exact time stamp does not exist, the iterator returned is for the next (later) record. If the time stamp is older than the current trim horizon, the iterator returned is for the oldest untrimmed data record (TRIM_HORIZON).

streamName :: Text

The name of the Amazon Kinesis data stream.

shardId :: Text

The shard ID of the Kinesis Data Streams shard to get the iterator for.

shardIteratorType :: ShardIteratorType

Determines how the shard iterator is used to start reading data records from the shard.

The following are the valid Amazon Kinesis shard iterator types:

AT_SEQUENCE_NUMBER - Start reading from the position denoted by a specific sequence number, provided in the value StartingSequenceNumber.

AFTER_SEQUENCE_NUMBER - Start reading right after the position denoted by a specific sequence number, provided in the value StartingSequenceNumber.

AT_TIMESTAMP - Start reading from the position denoted by a specific time stamp, provided in the value Timestamp.

TRIM_HORIZON - Start reading at the last untrimmed record in the shard in the system, which is the oldest data record in the shard.

LATEST - Start reading just after the most recent record in the shard, so that you always read the most recent data in the shard.

newGetShardIterator

Arguments

:: Text
$sel:streamName:GetShardIterator'

-> Text
$sel:shardId:GetShardIterator'

-> ShardIteratorType
$sel:shardIteratorType:GetShardIterator'

-> GetShardIterator

Create a value of GetShardIterator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startingSequenceNumber:GetShardIterator', getShardIterator_startingSequenceNumber - The sequence number of the data record in the shard from which to start reading. Used with shard iterator type AT_SEQUENCE_NUMBER and AFTER_SEQUENCE_NUMBER.

$sel:timestamp:GetShardIterator', getShardIterator_timestamp - The time stamp of the data record from which to start reading. Used with shard iterator type AT_TIMESTAMP. A time stamp is the Unix epoch date with precision in milliseconds. For example, 2016-04-04T19:58:46.480-00:00 or 1459799926.480. If a record with this exact time stamp does not exist, the iterator returned is for the next (later) record. If the time stamp is older than the current trim horizon, the iterator returned is for the oldest untrimmed data record (TRIM_HORIZON).

$sel:streamName:GetShardIterator', getShardIterator_streamName - The name of the Amazon Kinesis data stream.

$sel:shardId:GetShardIterator', getShardIterator_shardId - The shard ID of the Kinesis Data Streams shard to get the iterator for.

$sel:shardIteratorType:GetShardIterator', getShardIterator_shardIteratorType - Determines how the shard iterator is used to start reading data records from the shard.

The following are the valid Amazon Kinesis shard iterator types:

AT_SEQUENCE_NUMBER - Start reading from the position denoted by a specific sequence number, provided in the value StartingSequenceNumber.

AFTER_SEQUENCE_NUMBER - Start reading right after the position denoted by a specific sequence number, provided in the value StartingSequenceNumber.

AT_TIMESTAMP - Start reading from the position denoted by a specific time stamp, provided in the value Timestamp.

TRIM_HORIZON - Start reading at the last untrimmed record in the shard in the system, which is the oldest data record in the shard.

LATEST - Start reading just after the most recent record in the shard, so that you always read the most recent data in the shard.

Request Lenses

getShardIterator_startingSequenceNumber :: Lens' GetShardIterator (Maybe Text)

The sequence number of the data record in the shard from which to start reading. Used with shard iterator type AT_SEQUENCE_NUMBER and AFTER_SEQUENCE_NUMBER.

getShardIterator_timestamp :: Lens' GetShardIterator (Maybe UTCTime)

The time stamp of the data record from which to start reading. Used with shard iterator type AT_TIMESTAMP. A time stamp is the Unix epoch date with precision in milliseconds. For example, 2016-04-04T19:58:46.480-00:00 or 1459799926.480. If a record with this exact time stamp does not exist, the iterator returned is for the next (later) record. If the time stamp is older than the current trim horizon, the iterator returned is for the oldest untrimmed data record (TRIM_HORIZON).

getShardIterator_streamName :: Lens' GetShardIterator Text

The name of the Amazon Kinesis data stream.

getShardIterator_shardId :: Lens' GetShardIterator Text

The shard ID of the Kinesis Data Streams shard to get the iterator for.

getShardIterator_shardIteratorType :: Lens' GetShardIterator ShardIteratorType

Determines how the shard iterator is used to start reading data records from the shard.

The following are the valid Amazon Kinesis shard iterator types:

AT_SEQUENCE_NUMBER - Start reading from the position denoted by a specific sequence number, provided in the value StartingSequenceNumber.

AFTER_SEQUENCE_NUMBER - Start reading right after the position denoted by a specific sequence number, provided in the value StartingSequenceNumber.

AT_TIMESTAMP - Start reading from the position denoted by a specific time stamp, provided in the value Timestamp.

TRIM_HORIZON - Start reading at the last untrimmed record in the shard in the system, which is the oldest data record in the shard.

LATEST - Start reading just after the most recent record in the shard, so that you always read the most recent data in the shard.

Destructuring the Response

data GetShardIteratorResponse

Represents the output for GetShardIterator.

See: newGetShardIteratorResponse smart constructor.

Constructors

GetShardIteratorResponse'

Fields

shardIterator :: Maybe Text

The position in the shard from which to start reading data records sequentially. A shard iterator specifies this position using the sequence number of a data record in a shard.

httpStatus :: Int

The response's http status code.

newGetShardIteratorResponse

Arguments

:: Int
$sel:httpStatus:GetShardIteratorResponse'

-> GetShardIteratorResponse

Create a value of GetShardIteratorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shardIterator:GetShardIteratorResponse', getShardIteratorResponse_shardIterator - The position in the shard from which to start reading data records sequentially. A shard iterator specifies this position using the sequence number of a data record in a shard.

$sel:httpStatus:GetShardIteratorResponse', getShardIteratorResponse_httpStatus - The response's http status code.

Response Lenses

getShardIteratorResponse_shardIterator :: Lens' GetShardIteratorResponse (Maybe Text)

The position in the shard from which to start reading data records sequentially. A shard iterator specifies this position using the sequence number of a data record in a shard.

getShardIteratorResponse_httpStatus :: Lens' GetShardIteratorResponse Int

The response's http status code.