Network.AWS.Kinesis.IncreaseStreamRetentionPeriod

Description

Increases the Kinesis data stream's retention period, which is the length of time data records are accessible after they are added to the stream. The maximum value of a stream's retention period is 168 hours (7 days).

If you choose a longer stream retention period, this operation increases the time period during which records that have not yet expired are accessible. However, it does not make previous, expired data (older than the stream's previous retention period) accessible after the operation has been called. For example, if a stream's retention period is set to 24 hours and is increased to 168 hours, any data that is older than 24 hours remains inaccessible to consumer applications.

Creating a Request

data IncreaseStreamRetentionPeriod

Represents the input for IncreaseStreamRetentionPeriod.

See: newIncreaseStreamRetentionPeriod smart constructor.

Constructors

IncreaseStreamRetentionPeriod'

Fields

streamName :: Text

The name of the stream to modify.

retentionPeriodHours :: Int

The new retention period of the stream, in hours. Must be more than the current retention period.

newIncreaseStreamRetentionPeriod

Arguments

:: Text
$sel:streamName:IncreaseStreamRetentionPeriod'

-> Int
$sel:retentionPeriodHours:IncreaseStreamRetentionPeriod'

-> IncreaseStreamRetentionPeriod

Create a value of IncreaseStreamRetentionPeriod with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamName:IncreaseStreamRetentionPeriod', increaseStreamRetentionPeriod_streamName - The name of the stream to modify.

$sel:retentionPeriodHours:IncreaseStreamRetentionPeriod', increaseStreamRetentionPeriod_retentionPeriodHours - The new retention period of the stream, in hours. Must be more than the current retention period.

Request Lenses

increaseStreamRetentionPeriod_streamName :: Lens' IncreaseStreamRetentionPeriod Text

The name of the stream to modify.

increaseStreamRetentionPeriod_retentionPeriodHours :: Lens' IncreaseStreamRetentionPeriod Int

The new retention period of the stream, in hours. Must be more than the current retention period.

Destructuring the Response

data IncreaseStreamRetentionPeriodResponse

See: newIncreaseStreamRetentionPeriodResponse smart constructor.

Constructors

IncreaseStreamRetentionPeriodResponse'

newIncreaseStreamRetentionPeriodResponse :: IncreaseStreamRetentionPeriodResponse

Create a value of IncreaseStreamRetentionPeriodResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.