Network.AWS.Kinesis.RegisterStreamConsumer

Description

Registers a consumer with a Kinesis data stream. When you use this operation, the consumer you register can then call SubscribeToShard to receive data from the stream using enhanced fan-out, at a rate of up to 2 MiB per second for every shard you subscribe to. This rate is unaffected by the total number of consumers that read from the same stream.

You can register up to 20 consumers per stream. A given consumer can only be registered with one stream at a time.

For an example of how to use this operations, see Enhanced Fan-Out Using the Kinesis Data Streams API.

The use of this operation has a limit of five transactions per second per account. Also, only 5 consumers can be created simultaneously. In other words, you cannot have more than 5 consumers in a CREATING status at the same time. Registering a 6th consumer while there are 5 in a CREATING status results in a LimitExceededException.

Creating a Request

data RegisterStreamConsumer

See: newRegisterStreamConsumer smart constructor.

Constructors

RegisterStreamConsumer'

Fields

streamARN :: Text

The ARN of the Kinesis data stream that you want to register the consumer with. For more info, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

consumerName :: Text

For a given Kinesis data stream, each consumer must have a unique name. However, consumer names don't have to be unique across data streams.

newRegisterStreamConsumer

Arguments

:: Text
$sel:streamARN:RegisterStreamConsumer'

-> Text
$sel:consumerName:RegisterStreamConsumer'

-> RegisterStreamConsumer

Create a value of RegisterStreamConsumer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamARN:RegisterStreamConsumer', registerStreamConsumer_streamARN - The ARN of the Kinesis data stream that you want to register the consumer with. For more info, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

$sel:consumerName:RegisterStreamConsumer', registerStreamConsumer_consumerName - For a given Kinesis data stream, each consumer must have a unique name. However, consumer names don't have to be unique across data streams.

Request Lenses

registerStreamConsumer_streamARN :: Lens' RegisterStreamConsumer Text

The ARN of the Kinesis data stream that you want to register the consumer with. For more info, see Amazon Resource Names (ARNs) and AWS Service Namespaces.

registerStreamConsumer_consumerName :: Lens' RegisterStreamConsumer Text

For a given Kinesis data stream, each consumer must have a unique name. However, consumer names don't have to be unique across data streams.

Destructuring the Response

data RegisterStreamConsumerResponse

See: newRegisterStreamConsumerResponse smart constructor.

Constructors

RegisterStreamConsumerResponse'

Fields

httpStatus :: Int

The response's http status code.

consumer :: Consumer

An object that represents the details of the consumer you registered. When you register a consumer, it gets an ARN that is generated by Kinesis Data Streams.

newRegisterStreamConsumerResponse

Arguments

:: Int
$sel:httpStatus:RegisterStreamConsumerResponse'

-> Consumer
$sel:consumer:RegisterStreamConsumerResponse'

-> RegisterStreamConsumerResponse

Create a value of RegisterStreamConsumerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:RegisterStreamConsumerResponse', registerStreamConsumerResponse_httpStatus - The response's http status code.

$sel:consumer:RegisterStreamConsumerResponse', registerStreamConsumerResponse_consumer - An object that represents the details of the consumer you registered. When you register a consumer, it gets an ARN that is generated by Kinesis Data Streams.

Response Lenses

registerStreamConsumerResponse_httpStatus :: Lens' RegisterStreamConsumerResponse Int

The response's http status code.

registerStreamConsumerResponse_consumer :: Lens' RegisterStreamConsumerResponse Consumer

An object that represents the details of the consumer you registered. When you register a consumer, it gets an ARN that is generated by Kinesis Data Streams.