Network.AWS.Kinesis.Types.ConsumerDescription

Description

Documentation

data ConsumerDescription

An object that represents the details of a registered consumer. This type of object is returned by DescribeStreamConsumer.

See: newConsumerDescription smart constructor.

Constructors

ConsumerDescription'

Fields

consumerName :: Text

The name of the consumer is something you choose when you register the consumer.

consumerARN :: Text

When you register a consumer, Kinesis Data Streams generates an ARN for it. You need this ARN to be able to call SubscribeToShard.

If you delete a consumer and then create a new one with the same name, it won't have the same ARN. That's because consumer ARNs contain the creation timestamp. This is important to keep in mind if you have IAM policies that reference consumer ARNs.

consumerStatus :: ConsumerStatus

A consumer can't read data while in the CREATING or DELETING states.

consumerCreationTimestamp :: POSIX

streamARN :: Text

The ARN of the stream with which you registered the consumer.

newConsumerDescription

Arguments

:: Text
$sel:consumerName:ConsumerDescription'

-> Text
$sel:consumerARN:ConsumerDescription'

-> ConsumerStatus
$sel:consumerStatus:ConsumerDescription'

-> UTCTime
$sel:consumerCreationTimestamp:ConsumerDescription'

-> Text
$sel:streamARN:ConsumerDescription'

-> ConsumerDescription

Create a value of ConsumerDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:consumerName:ConsumerDescription', consumerDescription_consumerName - The name of the consumer is something you choose when you register the consumer.

$sel:consumerARN:ConsumerDescription', consumerDescription_consumerARN - When you register a consumer, Kinesis Data Streams generates an ARN for it. You need this ARN to be able to call SubscribeToShard.

If you delete a consumer and then create a new one with the same name, it won't have the same ARN. That's because consumer ARNs contain the creation timestamp. This is important to keep in mind if you have IAM policies that reference consumer ARNs.

$sel:consumerStatus:ConsumerDescription', consumerDescription_consumerStatus - A consumer can't read data while in the CREATING or DELETING states.

$sel:consumerCreationTimestamp:ConsumerDescription', consumerDescription_consumerCreationTimestamp -

$sel:streamARN:ConsumerDescription', consumerDescription_streamARN - The ARN of the stream with which you registered the consumer.

consumerDescription_consumerName :: Lens' ConsumerDescription Text

The name of the consumer is something you choose when you register the consumer.

consumerDescription_consumerARN :: Lens' ConsumerDescription Text

When you register a consumer, Kinesis Data Streams generates an ARN for it. You need this ARN to be able to call SubscribeToShard.

If you delete a consumer and then create a new one with the same name, it won't have the same ARN. That's because consumer ARNs contain the creation timestamp. This is important to keep in mind if you have IAM policies that reference consumer ARNs.

consumerDescription_consumerStatus :: Lens' ConsumerDescription ConsumerStatus

A consumer can't read data while in the CREATING or DELETING states.

consumerDescription_consumerCreationTimestamp :: Lens' ConsumerDescription UTCTime

consumerDescription_streamARN :: Lens' ConsumerDescription Text

The ARN of the stream with which you registered the consumer.