Network.AWS.Kinesis.Types.PutRecordsResultEntry

Description

Documentation

data PutRecordsResultEntry

Represents the result of an individual record from a PutRecords request. A record that is successfully added to a stream includes SequenceNumber and ShardId in the result. A record that fails to be added to the stream includes ErrorCode and ErrorMessage in the result.

See: newPutRecordsResultEntry smart constructor.

Constructors

PutRecordsResultEntry'

Fields

sequenceNumber :: Maybe Text

The sequence number for an individual record result.

errorCode :: Maybe Text

The error code for an individual record result. ErrorCodes can be either ProvisionedThroughputExceededException or InternalFailure.

errorMessage :: Maybe Text

The error message for an individual record result. An ErrorCode value of ProvisionedThroughputExceededException has an error message that includes the account ID, stream name, and shard ID. An ErrorCode value of InternalFailure has the error message "Internal Service Failure".

shardId :: Maybe Text

The shard ID for an individual record result.

newPutRecordsResultEntry :: PutRecordsResultEntry

Create a value of PutRecordsResultEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sequenceNumber:PutRecordsResultEntry', putRecordsResultEntry_sequenceNumber - The sequence number for an individual record result.

$sel:errorCode:PutRecordsResultEntry', putRecordsResultEntry_errorCode - The error code for an individual record result. ErrorCodes can be either ProvisionedThroughputExceededException or InternalFailure.

$sel:errorMessage:PutRecordsResultEntry', putRecordsResultEntry_errorMessage - The error message for an individual record result. An ErrorCode value of ProvisionedThroughputExceededException has an error message that includes the account ID, stream name, and shard ID. An ErrorCode value of InternalFailure has the error message "Internal Service Failure".

$sel:shardId:PutRecordsResultEntry', putRecordsResultEntry_shardId - The shard ID for an individual record result.

putRecordsResultEntry_sequenceNumber :: Lens' PutRecordsResultEntry (Maybe Text)

The sequence number for an individual record result.

putRecordsResultEntry_errorCode :: Lens' PutRecordsResultEntry (Maybe Text)

The error code for an individual record result. ErrorCodes can be either ProvisionedThroughputExceededException or InternalFailure.

putRecordsResultEntry_errorMessage :: Lens' PutRecordsResultEntry (Maybe Text)

The error message for an individual record result. An ErrorCode value of ProvisionedThroughputExceededException has an error message that includes the account ID, stream name, and shard ID. An ErrorCode value of InternalFailure has the error message "Internal Service Failure".

putRecordsResultEntry_shardId :: Lens' PutRecordsResultEntry (Maybe Text)

The shard ID for an individual record result.