Network.AWS.Kinesis.Types.StreamDescription

Description

Documentation

data StreamDescription

Represents the output for DescribeStream.

See: newStreamDescription smart constructor.

Constructors

StreamDescription'

Fields

encryptionType :: Maybe EncryptionType

The server-side encryption type used on the stream. This parameter can be one of the following values:

NONE: Do not encrypt the records in the stream.

KMS: Use server-side encryption on the records in the stream using a customer-managed AWS KMS key.

keyId :: Maybe Text

The GUID for the customer-managed AWS KMS key to use for encryption. This value can be a globally unique identifier, a fully specified ARN to either an alias or a key, or an alias name prefixed by "alias/".You can also use a master key owned by Kinesis Data Streams by specifying the alias aws/kinesis.

Key ARN example: arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

Alias ARN example: arn:aws:kms:us-east-1:123456789012:alias/MyAliasName

Globally unique key ID example: 12345678-1234-1234-1234-123456789012

Alias name example: alias/MyAliasName

Master key owned by Kinesis Data Streams: alias/aws/kinesis

streamName :: Text

The name of the stream being described.

streamARN :: Text

The Amazon Resource Name (ARN) for the stream being described.

streamStatus :: StreamStatus

The current status of the stream being described. The stream status is one of the following states:

CREATING - The stream is being created. Kinesis Data Streams immediately returns and sets StreamStatus to CREATING.

DELETING - The stream is being deleted. The specified stream is in the DELETING state until Kinesis Data Streams completes the deletion.

ACTIVE - The stream exists and is ready for read and write operations or deletion. You should perform read and write operations only on an ACTIVE stream.

UPDATING - Shards in the stream are being merged or split. Read and write operations continue to work while the stream is in the UPDATING state.

shards :: [Shard]

The shards that comprise the stream.

hasMoreShards :: Bool

If set to true, more shards in the stream are available to describe.

retentionPeriodHours :: Int

The current retention period, in hours. Minimum value of 24. Maximum value of 168.

streamCreationTimestamp :: POSIX

The approximate time that the stream was created.

enhancedMonitoring :: [EnhancedMetrics]

Represents the current enhanced monitoring settings of the stream.

newStreamDescription

Arguments

:: Text
$sel:streamName:StreamDescription'

-> Text
$sel:streamARN:StreamDescription'

-> StreamStatus
$sel:streamStatus:StreamDescription'

-> Bool
$sel:hasMoreShards:StreamDescription'

-> Int
$sel:retentionPeriodHours:StreamDescription'

-> UTCTime
$sel:streamCreationTimestamp:StreamDescription'

-> StreamDescription

Create a value of StreamDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:encryptionType:StreamDescription', streamDescription_encryptionType - The server-side encryption type used on the stream. This parameter can be one of the following values:

NONE: Do not encrypt the records in the stream.

KMS: Use server-side encryption on the records in the stream using a customer-managed AWS KMS key.

$sel:keyId:StreamDescription', streamDescription_keyId - The GUID for the customer-managed AWS KMS key to use for encryption. This value can be a globally unique identifier, a fully specified ARN to either an alias or a key, or an alias name prefixed by "alias/".You can also use a master key owned by Kinesis Data Streams by specifying the alias aws/kinesis.

Key ARN example: arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

Alias ARN example: arn:aws:kms:us-east-1:123456789012:alias/MyAliasName

Globally unique key ID example: 12345678-1234-1234-1234-123456789012

Alias name example: alias/MyAliasName

Master key owned by Kinesis Data Streams: alias/aws/kinesis

$sel:streamName:StreamDescription', streamDescription_streamName - The name of the stream being described.

$sel:streamARN:StreamDescription', streamDescription_streamARN - The Amazon Resource Name (ARN) for the stream being described.

$sel:streamStatus:StreamDescription', streamDescription_streamStatus - The current status of the stream being described. The stream status is one of the following states:

CREATING - The stream is being created. Kinesis Data Streams immediately returns and sets StreamStatus to CREATING.

DELETING - The stream is being deleted. The specified stream is in the DELETING state until Kinesis Data Streams completes the deletion.

ACTIVE - The stream exists and is ready for read and write operations or deletion. You should perform read and write operations only on an ACTIVE stream.

UPDATING - Shards in the stream are being merged or split. Read and write operations continue to work while the stream is in the UPDATING state.

$sel:shards:StreamDescription', streamDescription_shards - The shards that comprise the stream.

$sel:hasMoreShards:StreamDescription', streamDescription_hasMoreShards - If set to true, more shards in the stream are available to describe.

$sel:retentionPeriodHours:StreamDescription', streamDescription_retentionPeriodHours - The current retention period, in hours. Minimum value of 24. Maximum value of 168.

$sel:streamCreationTimestamp:StreamDescription', streamDescription_streamCreationTimestamp - The approximate time that the stream was created.

$sel:enhancedMonitoring:StreamDescription', streamDescription_enhancedMonitoring - Represents the current enhanced monitoring settings of the stream.

streamDescription_encryptionType :: Lens' StreamDescription (Maybe EncryptionType)

The server-side encryption type used on the stream. This parameter can be one of the following values:

NONE: Do not encrypt the records in the stream.

KMS: Use server-side encryption on the records in the stream using a customer-managed AWS KMS key.

streamDescription_keyId :: Lens' StreamDescription (Maybe Text)

The GUID for the customer-managed AWS KMS key to use for encryption. This value can be a globally unique identifier, a fully specified ARN to either an alias or a key, or an alias name prefixed by "alias/".You can also use a master key owned by Kinesis Data Streams by specifying the alias aws/kinesis.

Key ARN example: arn:aws:kms:us-east-1:123456789012:key/12345678-1234-1234-1234-123456789012

Alias ARN example: arn:aws:kms:us-east-1:123456789012:alias/MyAliasName

Globally unique key ID example: 12345678-1234-1234-1234-123456789012

Alias name example: alias/MyAliasName

Master key owned by Kinesis Data Streams: alias/aws/kinesis

streamDescription_streamName :: Lens' StreamDescription Text

The name of the stream being described.

streamDescription_streamARN :: Lens' StreamDescription Text

The Amazon Resource Name (ARN) for the stream being described.

streamDescription_streamStatus :: Lens' StreamDescription StreamStatus

The current status of the stream being described. The stream status is one of the following states:

CREATING - The stream is being created. Kinesis Data Streams immediately returns and sets StreamStatus to CREATING.

DELETING - The stream is being deleted. The specified stream is in the DELETING state until Kinesis Data Streams completes the deletion.

ACTIVE - The stream exists and is ready for read and write operations or deletion. You should perform read and write operations only on an ACTIVE stream.

UPDATING - Shards in the stream are being merged or split. Read and write operations continue to work while the stream is in the UPDATING state.

streamDescription_shards :: Lens' StreamDescription [Shard]

The shards that comprise the stream.

streamDescription_hasMoreShards :: Lens' StreamDescription Bool

If set to true, more shards in the stream are available to describe.

streamDescription_retentionPeriodHours :: Lens' StreamDescription Int

The current retention period, in hours. Minimum value of 24. Maximum value of 168.

streamDescription_streamCreationTimestamp :: Lens' StreamDescription UTCTime

The approximate time that the stream was created.

streamDescription_enhancedMonitoring :: Lens' StreamDescription [EnhancedMetrics]

Represents the current enhanced monitoring settings of the stream.