Network.AWS.Kinesis.Types.SubscribeToShardEvent

Description

Documentation

data SubscribeToShardEvent

After you call SubscribeToShard, Kinesis Data Streams sends events of this type over an HTTP/2 connection to your consumer.

See: newSubscribeToShardEvent smart constructor.

Constructors

SubscribeToShardEvent'

Fields

childShards :: Maybe [ChildShard]

records :: [Record]

continuationSequenceNumber :: Text

Use this as SequenceNumber in the next call to SubscribeToShard, with StartingPosition set to AT_SEQUENCE_NUMBER or AFTER_SEQUENCE_NUMBER. Use ContinuationSequenceNumber for checkpointing because it captures your shard progress even when no data is written to the shard.

millisBehindLatest :: Natural

The number of milliseconds the read records are from the tip of the stream, indicating how far behind current time the consumer is. A value of zero indicates that record processing is caught up, and there are no new records to process at this moment.

newSubscribeToShardEvent

Arguments

:: Text
$sel:continuationSequenceNumber:SubscribeToShardEvent'

-> Natural
$sel:millisBehindLatest:SubscribeToShardEvent'

-> SubscribeToShardEvent

Create a value of SubscribeToShardEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:childShards:SubscribeToShardEvent', subscribeToShardEvent_childShards - Undocumented member.

$sel:records:SubscribeToShardEvent', subscribeToShardEvent_records -

$sel:continuationSequenceNumber:SubscribeToShardEvent', subscribeToShardEvent_continuationSequenceNumber - Use this as SequenceNumber in the next call to SubscribeToShard, with StartingPosition set to AT_SEQUENCE_NUMBER or AFTER_SEQUENCE_NUMBER. Use ContinuationSequenceNumber for checkpointing because it captures your shard progress even when no data is written to the shard.

$sel:millisBehindLatest:SubscribeToShardEvent', subscribeToShardEvent_millisBehindLatest - The number of milliseconds the read records are from the tip of the stream, indicating how far behind current time the consumer is. A value of zero indicates that record processing is caught up, and there are no new records to process at this moment.

subscribeToShardEvent_childShards :: Lens' SubscribeToShardEvent (Maybe [ChildShard])

Undocumented member.

subscribeToShardEvent_records :: Lens' SubscribeToShardEvent [Record]

subscribeToShardEvent_continuationSequenceNumber :: Lens' SubscribeToShardEvent Text

Use this as SequenceNumber in the next call to SubscribeToShard, with StartingPosition set to AT_SEQUENCE_NUMBER or AFTER_SEQUENCE_NUMBER. Use ContinuationSequenceNumber for checkpointing because it captures your shard progress even when no data is written to the shard.

subscribeToShardEvent_millisBehindLatest :: Lens' SubscribeToShardEvent Natural

The number of milliseconds the read records are from the tip of the stream, indicating how far behind current time the consumer is. A value of zero indicates that record processing is caught up, and there are no new records to process at this moment.