Network.AWS.KinesisAnalytics.DescribeApplication

Description

This documentation is for version 1 of the Amazon Kinesis Data Analytics API, which only supports SQL applications. Version 2 of the API supports SQL and Java applications. For more information about version 2, see Amazon Kinesis Data Analytics API V2 Documentation.

Returns information about a specific Amazon Kinesis Analytics application.

If you want to retrieve a list of all applications in your account, use the ListApplications operation.

This operation requires permissions to perform the kinesisanalytics:DescribeApplication action. You can use DescribeApplication to get the current application versionId, which you need to call other operations such as Update.

Creating a Request

data DescribeApplication

See: newDescribeApplication smart constructor.

Constructors

DescribeApplication'

Fields

applicationName :: Text

Name of the application.

newDescribeApplication

Arguments

:: Text
$sel:applicationName:DescribeApplication'

-> DescribeApplication

Create a value of DescribeApplication with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationName:DescribeApplication', describeApplication_applicationName - Name of the application.

Request Lenses

describeApplication_applicationName :: Lens' DescribeApplication Text

Name of the application.

Destructuring the Response

data DescribeApplicationResponse

See: newDescribeApplicationResponse smart constructor.

Constructors

DescribeApplicationResponse'

Fields

httpStatus :: Int

The response's http status code.

applicationDetail :: ApplicationDetail

Provides a description of the application, such as the application Amazon Resource Name (ARN), status, latest version, and input and output configuration details.

newDescribeApplicationResponse

Arguments

:: Int
$sel:httpStatus:DescribeApplicationResponse'

-> ApplicationDetail
$sel:applicationDetail:DescribeApplicationResponse'

-> DescribeApplicationResponse

Create a value of DescribeApplicationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeApplicationResponse', describeApplicationResponse_httpStatus - The response's http status code.

$sel:applicationDetail:DescribeApplicationResponse', describeApplicationResponse_applicationDetail - Provides a description of the application, such as the application Amazon Resource Name (ARN), status, latest version, and input and output configuration details.

Response Lenses

describeApplicationResponse_httpStatus :: Lens' DescribeApplicationResponse Int

The response's http status code.

describeApplicationResponse_applicationDetail :: Lens' DescribeApplicationResponse ApplicationDetail

Provides a description of the application, such as the application Amazon Resource Name (ARN), status, latest version, and input and output configuration details.