Network.AWS.KinesisAnalytics.DiscoverInputSchema

Description

This documentation is for version 1 of the Amazon Kinesis Data Analytics API, which only supports SQL applications. Version 2 of the API supports SQL and Java applications. For more information about version 2, see Amazon Kinesis Data Analytics API V2 Documentation.

Infers a schema by evaluating sample records on the specified streaming source (Amazon Kinesis stream or Amazon Kinesis Firehose delivery stream) or S3 object. In the response, the operation returns the inferred schema and also the sample records that the operation used to infer the schema.

You can use the inferred schema when configuring a streaming source for your application. For conceptual information, see Configuring Application Input. Note that when you create an application using the Amazon Kinesis Analytics console, the console uses this operation to infer a schema and show it in the console user interface.

This operation requires permissions to perform the kinesisanalytics:DiscoverInputSchema action.

Creating a Request

data DiscoverInputSchema

See: newDiscoverInputSchema smart constructor.

Constructors

DiscoverInputSchema'

Fields

inputStartingPositionConfiguration :: Maybe InputStartingPositionConfiguration

Point at which you want Amazon Kinesis Analytics to start reading records from the specified streaming source discovery purposes.

resourceARN :: Maybe Text

Amazon Resource Name (ARN) of the streaming source.

roleARN :: Maybe Text

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf.

s3Configuration :: Maybe S3Configuration

Specify this parameter to discover a schema from data in an Amazon S3 object.

inputProcessingConfiguration :: Maybe InputProcessingConfiguration

The InputProcessingConfiguration to use to preprocess the records before discovering the schema of the records.

newDiscoverInputSchema :: DiscoverInputSchema

Create a value of DiscoverInputSchema with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputStartingPositionConfiguration:DiscoverInputSchema', discoverInputSchema_inputStartingPositionConfiguration - Point at which you want Amazon Kinesis Analytics to start reading records from the specified streaming source discovery purposes.

$sel:resourceARN:DiscoverInputSchema', discoverInputSchema_resourceARN - Amazon Resource Name (ARN) of the streaming source.

$sel:roleARN:DiscoverInputSchema', discoverInputSchema_roleARN - ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf.

$sel:s3Configuration:DiscoverInputSchema', discoverInputSchema_s3Configuration - Specify this parameter to discover a schema from data in an Amazon S3 object.

$sel:inputProcessingConfiguration:DiscoverInputSchema', discoverInputSchema_inputProcessingConfiguration - The InputProcessingConfiguration to use to preprocess the records before discovering the schema of the records.

Request Lenses

discoverInputSchema_inputStartingPositionConfiguration :: Lens' DiscoverInputSchema (Maybe InputStartingPositionConfiguration)

Point at which you want Amazon Kinesis Analytics to start reading records from the specified streaming source discovery purposes.

discoverInputSchema_resourceARN :: Lens' DiscoverInputSchema (Maybe Text)

Amazon Resource Name (ARN) of the streaming source.

discoverInputSchema_roleARN :: Lens' DiscoverInputSchema (Maybe Text)

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf.

discoverInputSchema_s3Configuration :: Lens' DiscoverInputSchema (Maybe S3Configuration)

Specify this parameter to discover a schema from data in an Amazon S3 object.

discoverInputSchema_inputProcessingConfiguration :: Lens' DiscoverInputSchema (Maybe InputProcessingConfiguration)

The InputProcessingConfiguration to use to preprocess the records before discovering the schema of the records.

Destructuring the Response

data DiscoverInputSchemaResponse

See: newDiscoverInputSchemaResponse smart constructor.

Constructors

DiscoverInputSchemaResponse'

Fields

inputSchema :: Maybe SourceSchema

Schema inferred from the streaming source. It identifies the format of the data in the streaming source and how each data element maps to corresponding columns in the in-application stream that you can create.

processedInputRecords :: Maybe [Text]

Stream data that was modified by the processor specified in the InputProcessingConfiguration parameter.

rawInputRecords :: Maybe [Text]

Raw stream data that was sampled to infer the schema.

parsedInputRecords :: Maybe [[Text]]

An array of elements, where each element corresponds to a row in a stream record (a stream record can have more than one row).

httpStatus :: Int

The response's http status code.

newDiscoverInputSchemaResponse

Arguments

:: Int
$sel:httpStatus:DiscoverInputSchemaResponse'

-> DiscoverInputSchemaResponse

Create a value of DiscoverInputSchemaResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputSchema:DiscoverInputSchemaResponse', discoverInputSchemaResponse_inputSchema - Schema inferred from the streaming source. It identifies the format of the data in the streaming source and how each data element maps to corresponding columns in the in-application stream that you can create.

$sel:processedInputRecords:DiscoverInputSchemaResponse', discoverInputSchemaResponse_processedInputRecords - Stream data that was modified by the processor specified in the InputProcessingConfiguration parameter.

$sel:rawInputRecords:DiscoverInputSchemaResponse', discoverInputSchemaResponse_rawInputRecords - Raw stream data that was sampled to infer the schema.

$sel:parsedInputRecords:DiscoverInputSchemaResponse', discoverInputSchemaResponse_parsedInputRecords - An array of elements, where each element corresponds to a row in a stream record (a stream record can have more than one row).

$sel:httpStatus:DiscoverInputSchemaResponse', discoverInputSchemaResponse_httpStatus - The response's http status code.

Response Lenses

discoverInputSchemaResponse_inputSchema :: Lens' DiscoverInputSchemaResponse (Maybe SourceSchema)

Schema inferred from the streaming source. It identifies the format of the data in the streaming source and how each data element maps to corresponding columns in the in-application stream that you can create.

discoverInputSchemaResponse_processedInputRecords :: Lens' DiscoverInputSchemaResponse (Maybe [Text])

Stream data that was modified by the processor specified in the InputProcessingConfiguration parameter.

discoverInputSchemaResponse_rawInputRecords :: Lens' DiscoverInputSchemaResponse (Maybe [Text])

Raw stream data that was sampled to infer the schema.

discoverInputSchemaResponse_parsedInputRecords :: Lens' DiscoverInputSchemaResponse (Maybe [[Text]])

An array of elements, where each element corresponds to a row in a stream record (a stream record can have more than one row).

discoverInputSchemaResponse_httpStatus :: Lens' DiscoverInputSchemaResponse Int

The response's http status code.