Network.AWS.KinesisAnalytics.Lens

Description

Operations

DeleteApplicationOutput

deleteApplicationOutput_applicationName :: Lens' DeleteApplicationOutput Text

Amazon Kinesis Analytics application name.

deleteApplicationOutput_currentApplicationVersionId :: Lens' DeleteApplicationOutput Natural

Amazon Kinesis Analytics application version. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

deleteApplicationOutput_outputId :: Lens' DeleteApplicationOutput Text

The ID of the configuration to delete. Each output configuration that is added to the application, either when the application is created or later using the AddApplicationOutput operation, has a unique ID. You need to provide the ID to uniquely identify the output configuration that you want to delete from the application configuration. You can use the DescribeApplication operation to get the specific OutputId.

deleteApplicationOutputResponse_httpStatus :: Lens' DeleteApplicationOutputResponse Int

The response's http status code.

AddApplicationOutput

addApplicationOutput_applicationName :: Lens' AddApplicationOutput Text

Name of the application to which you want to add the output configuration.

addApplicationOutput_currentApplicationVersionId :: Lens' AddApplicationOutput Natural

Version of the application to which you want to add the output configuration. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

addApplicationOutput_output :: Lens' AddApplicationOutput Output

An array of objects, each describing one output configuration. In the output configuration, you specify the name of an in-application stream, a destination (that is, an Amazon Kinesis stream, an Amazon Kinesis Firehose delivery stream, or an AWS Lambda function), and record the formation to use when writing to the destination.

addApplicationOutputResponse_httpStatus :: Lens' AddApplicationOutputResponse Int

The response's http status code.

CreateApplication

createApplication_applicationCode :: Lens' CreateApplication (Maybe Text)

One or more SQL statements that read input data, transform it, and generate output. For example, you can write a SQL statement that reads data from one in-application stream, generates a running average of the number of advertisement clicks by vendor, and insert resulting rows in another in-application stream using pumps. For more information about the typical pattern, see Application Code.

You can provide such series of SQL statements, where output of one statement can be used as the input for the next statement. You store intermediate results by creating in-application streams and pumps.

Note that the application code must create the streams with names specified in the Outputs. For example, if your Outputs defines output streams named ExampleOutputStream1 and ExampleOutputStream2, then your application code must create these streams.

createApplication_applicationDescription :: Lens' CreateApplication (Maybe Text)

Summary description of the application.

createApplication_outputs :: Lens' CreateApplication (Maybe [Output])

You can configure application output to write data from any of the in-application streams to up to three destinations.

These destinations can be Amazon Kinesis streams, Amazon Kinesis Firehose delivery streams, AWS Lambda destinations, or any combination of the three.

In the configuration, you specify the in-application stream name, the destination stream or Lambda function Amazon Resource Name (ARN), and the format to use when writing data. You must also provide an IAM role that Amazon Kinesis Analytics can assume to write to the destination stream or Lambda function on your behalf.

In the output configuration, you also provide the output stream or Lambda function ARN. For stream destinations, you provide the format of data in the stream (for example, JSON, CSV). You also must provide an IAM role that Amazon Kinesis Analytics can assume to write to the stream or Lambda function on your behalf.

createApplication_cloudWatchLoggingOptions :: Lens' CreateApplication (Maybe [CloudWatchLoggingOption])

Use this parameter to configure a CloudWatch log stream to monitor application configuration errors. For more information, see Working with Amazon CloudWatch Logs.

createApplication_tags :: Lens' CreateApplication (Maybe (NonEmpty Tag))

A list of one or more tags to assign to the application. A tag is a key-value pair that identifies an application. Note that the maximum number of application tags includes system tags. The maximum number of user-defined application tags is 50. For more information, see Using Tagging.

createApplication_inputs :: Lens' CreateApplication (Maybe [Input])

Use this parameter to configure the application input.

You can configure your application to receive input from a single streaming source. In this configuration, you map this streaming source to an in-application stream that is created. Your application code can then query the in-application stream like a table (you can think of it as a constantly updating table).

For the streaming source, you provide its Amazon Resource Name (ARN) and format of data on the stream (for example, JSON, CSV, etc.). You also must provide an IAM role that Amazon Kinesis Analytics can assume to read this stream on your behalf.

To create the in-application stream, you need to specify a schema to transform your data into a schematized version used in SQL. In the schema, you provide the necessary mapping of the data elements in the streaming source to record columns in the in-app stream.

createApplication_applicationName :: Lens' CreateApplication Text

Name of your Amazon Kinesis Analytics application (for example, sample-app).

createApplicationResponse_httpStatus :: Lens' CreateApplicationResponse Int

The response's http status code.

createApplicationResponse_applicationSummary :: Lens' CreateApplicationResponse ApplicationSummary

In response to your CreateApplication request, Amazon Kinesis Analytics returns a response with a summary of the application it created, including the application Amazon Resource Name (ARN), name, and status.

DeleteApplicationCloudWatchLoggingOption

deleteApplicationCloudWatchLoggingOption_applicationName :: Lens' DeleteApplicationCloudWatchLoggingOption Text

The Kinesis Analytics application name.

deleteApplicationCloudWatchLoggingOption_currentApplicationVersionId :: Lens' DeleteApplicationCloudWatchLoggingOption Natural

The version ID of the Kinesis Analytics application.

deleteApplicationCloudWatchLoggingOption_cloudWatchLoggingOptionId :: Lens' DeleteApplicationCloudWatchLoggingOption Text

The CloudWatchLoggingOptionId of the CloudWatch logging option to delete. You can get the CloudWatchLoggingOptionId by using the DescribeApplication operation.

deleteApplicationCloudWatchLoggingOptionResponse_httpStatus :: Lens' DeleteApplicationCloudWatchLoggingOptionResponse Int

The response's http status code.

AddApplicationCloudWatchLoggingOption

addApplicationCloudWatchLoggingOption_applicationName :: Lens' AddApplicationCloudWatchLoggingOption Text

The Kinesis Analytics application name.

addApplicationCloudWatchLoggingOption_currentApplicationVersionId :: Lens' AddApplicationCloudWatchLoggingOption Natural

The version ID of the Kinesis Analytics application.

addApplicationCloudWatchLoggingOption_cloudWatchLoggingOption :: Lens' AddApplicationCloudWatchLoggingOption CloudWatchLoggingOption

Provides the CloudWatch log stream Amazon Resource Name (ARN) and the IAM role ARN. Note: To write application messages to CloudWatch, the IAM role that is used must have the PutLogEvents policy action enabled.

addApplicationCloudWatchLoggingOptionResponse_httpStatus :: Lens' AddApplicationCloudWatchLoggingOptionResponse Int

The response's http status code.

UntagResource

untagResource_resourceARN :: Lens' UntagResource Text

The ARN of the Kinesis Analytics application from which to remove the tags.

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

A list of keys of tags to remove from the specified application.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

TagResource

tagResource_resourceARN :: Lens' TagResource Text

The ARN of the application to assign the tags.

tagResource_tags :: Lens' TagResource (NonEmpty Tag)

The key-value tags to assign to the application.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

StartApplication

startApplication_applicationName :: Lens' StartApplication Text

Name of the application.

startApplication_inputConfigurations :: Lens' StartApplication [InputConfiguration]

Identifies the specific input, by ID, that the application starts consuming. Amazon Kinesis Analytics starts reading the streaming source associated with the input. You can also specify where in the streaming source you want Amazon Kinesis Analytics to start reading.

startApplicationResponse_httpStatus :: Lens' StartApplicationResponse Int

The response's http status code.

StopApplication

stopApplication_applicationName :: Lens' StopApplication Text

Name of the running application to stop.

stopApplicationResponse_httpStatus :: Lens' StopApplicationResponse Int

The response's http status code.

DescribeApplication

describeApplication_applicationName :: Lens' DescribeApplication Text

Name of the application.

describeApplicationResponse_httpStatus :: Lens' DescribeApplicationResponse Int

The response's http status code.

describeApplicationResponse_applicationDetail :: Lens' DescribeApplicationResponse ApplicationDetail

Provides a description of the application, such as the application Amazon Resource Name (ARN), status, latest version, and input and output configuration details.

DiscoverInputSchema

discoverInputSchema_inputStartingPositionConfiguration :: Lens' DiscoverInputSchema (Maybe InputStartingPositionConfiguration)

Point at which you want Amazon Kinesis Analytics to start reading records from the specified streaming source discovery purposes.

discoverInputSchema_resourceARN :: Lens' DiscoverInputSchema (Maybe Text)

Amazon Resource Name (ARN) of the streaming source.

discoverInputSchema_roleARN :: Lens' DiscoverInputSchema (Maybe Text)

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf.

discoverInputSchema_s3Configuration :: Lens' DiscoverInputSchema (Maybe S3Configuration)

Specify this parameter to discover a schema from data in an Amazon S3 object.

discoverInputSchema_inputProcessingConfiguration :: Lens' DiscoverInputSchema (Maybe InputProcessingConfiguration)

The InputProcessingConfiguration to use to preprocess the records before discovering the schema of the records.

discoverInputSchemaResponse_inputSchema :: Lens' DiscoverInputSchemaResponse (Maybe SourceSchema)

Schema inferred from the streaming source. It identifies the format of the data in the streaming source and how each data element maps to corresponding columns in the in-application stream that you can create.

discoverInputSchemaResponse_processedInputRecords :: Lens' DiscoverInputSchemaResponse (Maybe [Text])

Stream data that was modified by the processor specified in the InputProcessingConfiguration parameter.

discoverInputSchemaResponse_rawInputRecords :: Lens' DiscoverInputSchemaResponse (Maybe [Text])

Raw stream data that was sampled to infer the schema.

discoverInputSchemaResponse_parsedInputRecords :: Lens' DiscoverInputSchemaResponse (Maybe [[Text]])

An array of elements, where each element corresponds to a row in a stream record (a stream record can have more than one row).

discoverInputSchemaResponse_httpStatus :: Lens' DiscoverInputSchemaResponse Int

The response's http status code.

DeleteApplicationInputProcessingConfiguration

deleteApplicationInputProcessingConfiguration_applicationName :: Lens' DeleteApplicationInputProcessingConfiguration Text

The Kinesis Analytics application name.

deleteApplicationInputProcessingConfiguration_currentApplicationVersionId :: Lens' DeleteApplicationInputProcessingConfiguration Natural

The version ID of the Kinesis Analytics application.

deleteApplicationInputProcessingConfiguration_inputId :: Lens' DeleteApplicationInputProcessingConfiguration Text

The ID of the input configuration from which to delete the input processing configuration. You can get a list of the input IDs for an application by using the DescribeApplication operation.

deleteApplicationInputProcessingConfigurationResponse_httpStatus :: Lens' DeleteApplicationInputProcessingConfigurationResponse Int

The response's http status code.

AddApplicationInputProcessingConfiguration

addApplicationInputProcessingConfiguration_applicationName :: Lens' AddApplicationInputProcessingConfiguration Text

Name of the application to which you want to add the input processing configuration.

addApplicationInputProcessingConfiguration_currentApplicationVersionId :: Lens' AddApplicationInputProcessingConfiguration Natural

Version of the application to which you want to add the input processing configuration. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

addApplicationInputProcessingConfiguration_inputId :: Lens' AddApplicationInputProcessingConfiguration Text

The ID of the input configuration to add the input processing configuration to. You can get a list of the input IDs for an application using the DescribeApplication operation.

addApplicationInputProcessingConfiguration_inputProcessingConfiguration :: Lens' AddApplicationInputProcessingConfiguration InputProcessingConfiguration

The InputProcessingConfiguration to add to the application.

addApplicationInputProcessingConfigurationResponse_httpStatus :: Lens' AddApplicationInputProcessingConfigurationResponse Int

The response's http status code.

ListApplications

listApplications_exclusiveStartApplicationName :: Lens' ListApplications (Maybe Text)

Name of the application to start the list with. When using pagination to retrieve the list, you don't need to specify this parameter in the first request. However, in subsequent requests, you add the last application name from the previous response to get the next page of applications.

listApplications_limit :: Lens' ListApplications (Maybe Natural)

Maximum number of applications to list.

listApplicationsResponse_httpStatus :: Lens' ListApplicationsResponse Int

The response's http status code.

listApplicationsResponse_applicationSummaries :: Lens' ListApplicationsResponse [ApplicationSummary]

List of ApplicationSummary objects.

listApplicationsResponse_hasMoreApplications :: Lens' ListApplicationsResponse Bool

Returns true if there are more applications to retrieve.

DeleteApplication

deleteApplication_applicationName :: Lens' DeleteApplication Text

Name of the Amazon Kinesis Analytics application to delete.

deleteApplication_createTimestamp :: Lens' DeleteApplication UTCTime

You can use the DescribeApplication operation to get this value.

deleteApplicationResponse_httpStatus :: Lens' DeleteApplicationResponse Int

The response's http status code.

UpdateApplication

updateApplication_applicationName :: Lens' UpdateApplication Text

Name of the Amazon Kinesis Analytics application to update.

updateApplication_currentApplicationVersionId :: Lens' UpdateApplication Natural

The current application version ID. You can use the DescribeApplication operation to get this value.

updateApplication_applicationUpdate :: Lens' UpdateApplication ApplicationUpdate

Describes application updates.

updateApplicationResponse_httpStatus :: Lens' UpdateApplicationResponse Int

The response's http status code.

AddApplicationReferenceDataSource

addApplicationReferenceDataSource_applicationName :: Lens' AddApplicationReferenceDataSource Text

Name of an existing application.

addApplicationReferenceDataSource_currentApplicationVersionId :: Lens' AddApplicationReferenceDataSource Natural

Version of the application for which you are adding the reference data source. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

addApplicationReferenceDataSource_referenceDataSource :: Lens' AddApplicationReferenceDataSource ReferenceDataSource

The reference data source can be an object in your Amazon S3 bucket. Amazon Kinesis Analytics reads the object and copies the data into the in-application table that is created. You provide an S3 bucket, object key name, and the resulting in-application table that is created. You must also provide an IAM role with the necessary permissions that Amazon Kinesis Analytics can assume to read the object from your S3 bucket on your behalf.

addApplicationReferenceDataSourceResponse_httpStatus :: Lens' AddApplicationReferenceDataSourceResponse Int

The response's http status code.

DeleteApplicationReferenceDataSource

deleteApplicationReferenceDataSource_applicationName :: Lens' DeleteApplicationReferenceDataSource Text

Name of an existing application.

deleteApplicationReferenceDataSource_currentApplicationVersionId :: Lens' DeleteApplicationReferenceDataSource Natural

Version of the application. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

deleteApplicationReferenceDataSource_referenceId :: Lens' DeleteApplicationReferenceDataSource Text

ID of the reference data source. When you add a reference data source to your application using the AddApplicationReferenceDataSource, Amazon Kinesis Analytics assigns an ID. You can use the DescribeApplication operation to get the reference ID.

deleteApplicationReferenceDataSourceResponse_httpStatus :: Lens' DeleteApplicationReferenceDataSourceResponse Int

The response's http status code.

AddApplicationInput

addApplicationInput_applicationName :: Lens' AddApplicationInput Text

Name of your existing Amazon Kinesis Analytics application to which you want to add the streaming source.

addApplicationInput_currentApplicationVersionId :: Lens' AddApplicationInput Natural

Current version of your Amazon Kinesis Analytics application. You can use the DescribeApplication operation to find the current application version.

addApplicationInput_input :: Lens' AddApplicationInput Input

The Input to add.

addApplicationInputResponse_httpStatus :: Lens' AddApplicationInputResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceARN :: Lens' ListTagsForResource Text

The ARN of the application for which to retrieve tags.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (NonEmpty Tag))

The key-value tags assigned to the application.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

Types

ApplicationDetail

applicationDetail_applicationCode :: Lens' ApplicationDetail (Maybe Text)

Returns the application code that you provided to perform data analysis on any of the in-application streams in your application.

applicationDetail_applicationDescription :: Lens' ApplicationDetail (Maybe Text)

Description of the application.

applicationDetail_cloudWatchLoggingOptionDescriptions :: Lens' ApplicationDetail (Maybe [CloudWatchLoggingOptionDescription])

Describes the CloudWatch log streams that are configured to receive application messages. For more information about using CloudWatch log streams with Amazon Kinesis Analytics applications, see Working with Amazon CloudWatch Logs.

applicationDetail_outputDescriptions :: Lens' ApplicationDetail (Maybe [OutputDescription])

Describes the application output configuration. For more information, see Configuring Application Output.

applicationDetail_createTimestamp :: Lens' ApplicationDetail (Maybe UTCTime)

Time stamp when the application version was created.

applicationDetail_inputDescriptions :: Lens' ApplicationDetail (Maybe [InputDescription])

Describes the application input configuration. For more information, see Configuring Application Input.

applicationDetail_referenceDataSourceDescriptions :: Lens' ApplicationDetail (Maybe [ReferenceDataSourceDescription])

Describes reference data sources configured for the application. For more information, see Configuring Application Input.

applicationDetail_lastUpdateTimestamp :: Lens' ApplicationDetail (Maybe UTCTime)

Time stamp when the application was last updated.

applicationDetail_applicationName :: Lens' ApplicationDetail Text

Name of the application.

applicationDetail_applicationARN :: Lens' ApplicationDetail Text

ARN of the application.

applicationDetail_applicationStatus :: Lens' ApplicationDetail ApplicationStatus

Status of the application.

applicationDetail_applicationVersionId :: Lens' ApplicationDetail Natural

Provides the current application version.

ApplicationSummary

applicationSummary_applicationName :: Lens' ApplicationSummary Text

Name of the application.

applicationSummary_applicationARN :: Lens' ApplicationSummary Text

ARN of the application.

applicationSummary_applicationStatus :: Lens' ApplicationSummary ApplicationStatus

Status of the application.

ApplicationUpdate

applicationUpdate_inputUpdates :: Lens' ApplicationUpdate (Maybe [InputUpdate])

Describes application input configuration updates.

applicationUpdate_referenceDataSourceUpdates :: Lens' ApplicationUpdate (Maybe [ReferenceDataSourceUpdate])

Describes application reference data source updates.

applicationUpdate_cloudWatchLoggingOptionUpdates :: Lens' ApplicationUpdate (Maybe [CloudWatchLoggingOptionUpdate])

Describes application CloudWatch logging option updates.

applicationUpdate_applicationCodeUpdate :: Lens' ApplicationUpdate (Maybe Text)

Describes application code updates.

applicationUpdate_outputUpdates :: Lens' ApplicationUpdate (Maybe [OutputUpdate])

Describes application output configuration updates.

CSVMappingParameters

cSVMappingParameters_recordRowDelimiter :: Lens' CSVMappingParameters Text

Row delimiter. For example, in a CSV format, '\n' is the typical row delimiter.

cSVMappingParameters_recordColumnDelimiter :: Lens' CSVMappingParameters Text

Column delimiter. For example, in a CSV format, a comma (",") is the typical column delimiter.

CloudWatchLoggingOption

cloudWatchLoggingOption_logStreamARN :: Lens' CloudWatchLoggingOption Text

ARN of the CloudWatch log to receive application messages.

cloudWatchLoggingOption_roleARN :: Lens' CloudWatchLoggingOption Text

IAM ARN of the role to use to send application messages. Note: To write application messages to CloudWatch, the IAM role that is used must have the PutLogEvents policy action enabled.

CloudWatchLoggingOptionDescription

cloudWatchLoggingOptionDescription_cloudWatchLoggingOptionId :: Lens' CloudWatchLoggingOptionDescription (Maybe Text)

ID of the CloudWatch logging option description.

cloudWatchLoggingOptionDescription_logStreamARN :: Lens' CloudWatchLoggingOptionDescription Text

ARN of the CloudWatch log to receive application messages.

cloudWatchLoggingOptionDescription_roleARN :: Lens' CloudWatchLoggingOptionDescription Text

IAM ARN of the role to use to send application messages. Note: To write application messages to CloudWatch, the IAM role used must have the PutLogEvents policy action enabled.

CloudWatchLoggingOptionUpdate

cloudWatchLoggingOptionUpdate_logStreamARNUpdate :: Lens' CloudWatchLoggingOptionUpdate (Maybe Text)

ARN of the CloudWatch log to receive application messages.

cloudWatchLoggingOptionUpdate_roleARNUpdate :: Lens' CloudWatchLoggingOptionUpdate (Maybe Text)

IAM ARN of the role to use to send application messages. Note: To write application messages to CloudWatch, the IAM role used must have the PutLogEvents policy action enabled.

cloudWatchLoggingOptionUpdate_cloudWatchLoggingOptionId :: Lens' CloudWatchLoggingOptionUpdate Text

ID of the CloudWatch logging option to update

DestinationSchema

destinationSchema_recordFormatType :: Lens' DestinationSchema RecordFormatType

Specifies the format of the records on the output stream.

Input

input_inputParallelism :: Lens' Input (Maybe InputParallelism)

Describes the number of in-application streams to create.

Data from your source is routed to these in-application input streams.

(see Configuring Application Input.

input_kinesisFirehoseInput :: Lens' Input (Maybe KinesisFirehoseInput)

If the streaming source is an Amazon Kinesis Firehose delivery stream, identifies the delivery stream's ARN and an IAM role that enables Amazon Kinesis Analytics to access the stream on your behalf.

Note: Either KinesisStreamsInput or KinesisFirehoseInput is required.

input_kinesisStreamsInput :: Lens' Input (Maybe KinesisStreamsInput)

If the streaming source is an Amazon Kinesis stream, identifies the stream's Amazon Resource Name (ARN) and an IAM role that enables Amazon Kinesis Analytics to access the stream on your behalf.

Note: Either KinesisStreamsInput or KinesisFirehoseInput is required.

input_inputProcessingConfiguration :: Lens' Input (Maybe InputProcessingConfiguration)

The InputProcessingConfiguration for the input. An input processor transforms records as they are received from the stream, before the application's SQL code executes. Currently, the only input processing configuration available is InputLambdaProcessor.

input_namePrefix :: Lens' Input Text

Name prefix to use when creating an in-application stream. Suppose that you specify a prefix "MyInApplicationStream." Amazon Kinesis Analytics then creates one or more (as per the InputParallelism count you specified) in-application streams with names "MyInApplicationStream_001," "MyInApplicationStream_002," and so on.

input_inputSchema :: Lens' Input SourceSchema

Describes the format of the data in the streaming source, and how each data element maps to corresponding columns in the in-application stream that is being created.

Also used to describe the format of the reference data source.

InputConfiguration

inputConfiguration_id :: Lens' InputConfiguration Text

Input source ID. You can get this ID by calling the DescribeApplication operation.

inputConfiguration_inputStartingPositionConfiguration :: Lens' InputConfiguration InputStartingPositionConfiguration

Point at which you want the application to start processing records from the streaming source.

InputDescription

inputDescription_inputSchema :: Lens' InputDescription (Maybe SourceSchema)

Describes the format of the data in the streaming source, and how each data element maps to corresponding columns in the in-application stream that is being created.

inputDescription_inputStartingPositionConfiguration :: Lens' InputDescription (Maybe InputStartingPositionConfiguration)

Point at which the application is configured to read from the input stream.

inputDescription_inputProcessingConfigurationDescription :: Lens' InputDescription (Maybe InputProcessingConfigurationDescription)

The description of the preprocessor that executes on records in this input before the application's code is run.

inputDescription_inputParallelism :: Lens' InputDescription (Maybe InputParallelism)

Describes the configured parallelism (number of in-application streams mapped to the streaming source).

inputDescription_namePrefix :: Lens' InputDescription (Maybe Text)

In-application name prefix.

inputDescription_kinesisStreamsInputDescription :: Lens' InputDescription (Maybe KinesisStreamsInputDescription)

If an Amazon Kinesis stream is configured as streaming source, provides Amazon Kinesis stream's Amazon Resource Name (ARN) and an IAM role that enables Amazon Kinesis Analytics to access the stream on your behalf.

inputDescription_kinesisFirehoseInputDescription :: Lens' InputDescription (Maybe KinesisFirehoseInputDescription)

If an Amazon Kinesis Firehose delivery stream is configured as a streaming source, provides the delivery stream's ARN and an IAM role that enables Amazon Kinesis Analytics to access the stream on your behalf.

inputDescription_inAppStreamNames :: Lens' InputDescription (Maybe [Text])

Returns the in-application stream names that are mapped to the stream source.

inputDescription_inputId :: Lens' InputDescription (Maybe Text)

Input ID associated with the application input. This is the ID that Amazon Kinesis Analytics assigns to each input configuration you add to your application.

InputLambdaProcessor

inputLambdaProcessor_resourceARN :: Lens' InputLambdaProcessor Text

The ARN of the AWS Lambda function that operates on records in the stream.

To specify an earlier version of the Lambda function than the latest, include the Lambda function version in the Lambda function ARN. For more information about Lambda ARNs, see Example ARNs: AWS Lambda

inputLambdaProcessor_roleARN :: Lens' InputLambdaProcessor Text

The ARN of the IAM role that is used to access the AWS Lambda function.

InputLambdaProcessorDescription

inputLambdaProcessorDescription_resourceARN :: Lens' InputLambdaProcessorDescription (Maybe Text)

The ARN of the AWS Lambda function that is used to preprocess the records in the stream.

inputLambdaProcessorDescription_roleARN :: Lens' InputLambdaProcessorDescription (Maybe Text)

The ARN of the IAM role that is used to access the AWS Lambda function.

InputLambdaProcessorUpdate

inputLambdaProcessorUpdate_resourceARNUpdate :: Lens' InputLambdaProcessorUpdate (Maybe Text)

The Amazon Resource Name (ARN) of the new AWS Lambda function that is used to preprocess the records in the stream.

To specify an earlier version of the Lambda function than the latest, include the Lambda function version in the Lambda function ARN. For more information about Lambda ARNs, see Example ARNs: AWS Lambda

inputLambdaProcessorUpdate_roleARNUpdate :: Lens' InputLambdaProcessorUpdate (Maybe Text)

The ARN of the new IAM role that is used to access the AWS Lambda function.

InputParallelism

inputParallelism_count :: Lens' InputParallelism (Maybe Natural)

Number of in-application streams to create. For more information, see Limits.

InputParallelismUpdate

inputParallelismUpdate_countUpdate :: Lens' InputParallelismUpdate (Maybe Natural)

Number of in-application streams to create for the specified streaming source.

InputProcessingConfiguration

inputProcessingConfiguration_inputLambdaProcessor :: Lens' InputProcessingConfiguration InputLambdaProcessor

The InputLambdaProcessor that is used to preprocess the records in the stream before being processed by your application code.

InputProcessingConfigurationDescription

inputProcessingConfigurationDescription_inputLambdaProcessorDescription :: Lens' InputProcessingConfigurationDescription (Maybe InputLambdaProcessorDescription)

Provides configuration information about the associated InputLambdaProcessorDescription.

InputProcessingConfigurationUpdate

inputProcessingConfigurationUpdate_inputLambdaProcessorUpdate :: Lens' InputProcessingConfigurationUpdate InputLambdaProcessorUpdate

Provides update information for an InputLambdaProcessor.

InputSchemaUpdate

inputSchemaUpdate_recordFormatUpdate :: Lens' InputSchemaUpdate (Maybe RecordFormat)

Specifies the format of the records on the streaming source.

inputSchemaUpdate_recordColumnUpdates :: Lens' InputSchemaUpdate (Maybe (NonEmpty RecordColumn))

A list of RecordColumn objects. Each object describes the mapping of the streaming source element to the corresponding column in the in-application stream.

inputSchemaUpdate_recordEncodingUpdate :: Lens' InputSchemaUpdate (Maybe Text)

Specifies the encoding of the records in the streaming source. For example, UTF-8.

InputStartingPositionConfiguration

inputStartingPositionConfiguration_inputStartingPosition :: Lens' InputStartingPositionConfiguration (Maybe InputStartingPosition)

The starting position on the stream.

NOW - Start reading just after the most recent record in the stream, start at the request time stamp that the customer issued.

TRIM_HORIZON - Start reading at the last untrimmed record in the stream, which is the oldest record available in the stream. This option is not available for an Amazon Kinesis Firehose delivery stream.

LAST_STOPPED_POINT - Resume reading from where the application last stopped reading.

InputUpdate

inputUpdate_kinesisFirehoseInputUpdate :: Lens' InputUpdate (Maybe KinesisFirehoseInputUpdate)

If an Amazon Kinesis Firehose delivery stream is the streaming source to be updated, provides an updated stream ARN and IAM role ARN.

inputUpdate_namePrefixUpdate :: Lens' InputUpdate (Maybe Text)

Name prefix for in-application streams that Amazon Kinesis Analytics creates for the specific streaming source.

inputUpdate_kinesisStreamsInputUpdate :: Lens' InputUpdate (Maybe KinesisStreamsInputUpdate)

If an Amazon Kinesis stream is the streaming source to be updated, provides an updated stream Amazon Resource Name (ARN) and IAM role ARN.

inputUpdate_inputProcessingConfigurationUpdate :: Lens' InputUpdate (Maybe InputProcessingConfigurationUpdate)

Describes updates for an input processing configuration.

inputUpdate_inputSchemaUpdate :: Lens' InputUpdate (Maybe InputSchemaUpdate)

Describes the data format on the streaming source, and how record elements on the streaming source map to columns of the in-application stream that is created.

inputUpdate_inputParallelismUpdate :: Lens' InputUpdate (Maybe InputParallelismUpdate)

Describes the parallelism updates (the number in-application streams Amazon Kinesis Analytics creates for the specific streaming source).

inputUpdate_inputId :: Lens' InputUpdate Text

Input ID of the application input to be updated.

JSONMappingParameters

jSONMappingParameters_recordRowPath :: Lens' JSONMappingParameters Text

Path to the top-level parent that contains the records.

KinesisFirehoseInput

kinesisFirehoseInput_resourceARN :: Lens' KinesisFirehoseInput Text

ARN of the input delivery stream.

kinesisFirehoseInput_roleARN :: Lens' KinesisFirehoseInput Text

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf. You need to make sure that the role has the necessary permissions to access the stream.

KinesisFirehoseInputDescription

kinesisFirehoseInputDescription_resourceARN :: Lens' KinesisFirehoseInputDescription (Maybe Text)

Amazon Resource Name (ARN) of the Amazon Kinesis Firehose delivery stream.

kinesisFirehoseInputDescription_roleARN :: Lens' KinesisFirehoseInputDescription (Maybe Text)

ARN of the IAM role that Amazon Kinesis Analytics assumes to access the stream.

KinesisFirehoseInputUpdate

kinesisFirehoseInputUpdate_resourceARNUpdate :: Lens' KinesisFirehoseInputUpdate (Maybe Text)

Amazon Resource Name (ARN) of the input Amazon Kinesis Firehose delivery stream to read.

kinesisFirehoseInputUpdate_roleARNUpdate :: Lens' KinesisFirehoseInputUpdate (Maybe Text)

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf. You need to grant the necessary permissions to this role.

KinesisFirehoseOutput

kinesisFirehoseOutput_resourceARN :: Lens' KinesisFirehoseOutput Text

ARN of the destination Amazon Kinesis Firehose delivery stream to write to.

kinesisFirehoseOutput_roleARN :: Lens' KinesisFirehoseOutput Text

ARN of the IAM role that Amazon Kinesis Analytics can assume to write to the destination stream on your behalf. You need to grant the necessary permissions to this role.

KinesisFirehoseOutputDescription

kinesisFirehoseOutputDescription_resourceARN :: Lens' KinesisFirehoseOutputDescription (Maybe Text)

Amazon Resource Name (ARN) of the Amazon Kinesis Firehose delivery stream.

kinesisFirehoseOutputDescription_roleARN :: Lens' KinesisFirehoseOutputDescription (Maybe Text)

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream.

KinesisFirehoseOutputUpdate

kinesisFirehoseOutputUpdate_resourceARNUpdate :: Lens' KinesisFirehoseOutputUpdate (Maybe Text)

Amazon Resource Name (ARN) of the Amazon Kinesis Firehose delivery stream to write to.

kinesisFirehoseOutputUpdate_roleARNUpdate :: Lens' KinesisFirehoseOutputUpdate (Maybe Text)

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf. You need to grant the necessary permissions to this role.

KinesisStreamsInput

kinesisStreamsInput_resourceARN :: Lens' KinesisStreamsInput Text

ARN of the input Amazon Kinesis stream to read.

kinesisStreamsInput_roleARN :: Lens' KinesisStreamsInput Text

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf. You need to grant the necessary permissions to this role.

KinesisStreamsInputDescription

kinesisStreamsInputDescription_resourceARN :: Lens' KinesisStreamsInputDescription (Maybe Text)

Amazon Resource Name (ARN) of the Amazon Kinesis stream.

kinesisStreamsInputDescription_roleARN :: Lens' KinesisStreamsInputDescription (Maybe Text)

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream.

KinesisStreamsInputUpdate

kinesisStreamsInputUpdate_resourceARNUpdate :: Lens' KinesisStreamsInputUpdate (Maybe Text)

Amazon Resource Name (ARN) of the input Amazon Kinesis stream to read.

kinesisStreamsInputUpdate_roleARNUpdate :: Lens' KinesisStreamsInputUpdate (Maybe Text)

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf. You need to grant the necessary permissions to this role.

KinesisStreamsOutput

kinesisStreamsOutput_resourceARN :: Lens' KinesisStreamsOutput Text

ARN of the destination Amazon Kinesis stream to write to.

kinesisStreamsOutput_roleARN :: Lens' KinesisStreamsOutput Text

ARN of the IAM role that Amazon Kinesis Analytics can assume to write to the destination stream on your behalf. You need to grant the necessary permissions to this role.

KinesisStreamsOutputDescription

kinesisStreamsOutputDescription_resourceARN :: Lens' KinesisStreamsOutputDescription (Maybe Text)

Amazon Resource Name (ARN) of the Amazon Kinesis stream.

kinesisStreamsOutputDescription_roleARN :: Lens' KinesisStreamsOutputDescription (Maybe Text)

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream.

KinesisStreamsOutputUpdate

kinesisStreamsOutputUpdate_resourceARNUpdate :: Lens' KinesisStreamsOutputUpdate (Maybe Text)

Amazon Resource Name (ARN) of the Amazon Kinesis stream where you want to write the output.

kinesisStreamsOutputUpdate_roleARNUpdate :: Lens' KinesisStreamsOutputUpdate (Maybe Text)

ARN of the IAM role that Amazon Kinesis Analytics can assume to access the stream on your behalf. You need to grant the necessary permissions to this role.

LambdaOutput

lambdaOutput_resourceARN :: Lens' LambdaOutput Text

Amazon Resource Name (ARN) of the destination Lambda function to write to.

To specify an earlier version of the Lambda function than the latest, include the Lambda function version in the Lambda function ARN. For more information about Lambda ARNs, see Example ARNs: AWS Lambda

lambdaOutput_roleARN :: Lens' LambdaOutput Text

ARN of the IAM role that Amazon Kinesis Analytics can assume to write to the destination function on your behalf. You need to grant the necessary permissions to this role.

LambdaOutputDescription

lambdaOutputDescription_resourceARN :: Lens' LambdaOutputDescription (Maybe Text)

Amazon Resource Name (ARN) of the destination Lambda function.

lambdaOutputDescription_roleARN :: Lens' LambdaOutputDescription (Maybe Text)

ARN of the IAM role that Amazon Kinesis Analytics can assume to write to the destination function.

LambdaOutputUpdate

lambdaOutputUpdate_resourceARNUpdate :: Lens' LambdaOutputUpdate (Maybe Text)

Amazon Resource Name (ARN) of the destination Lambda function.

To specify an earlier version of the Lambda function than the latest, include the Lambda function version in the Lambda function ARN. For more information about Lambda ARNs, see Example ARNs: AWS Lambda

lambdaOutputUpdate_roleARNUpdate :: Lens' LambdaOutputUpdate (Maybe Text)

ARN of the IAM role that Amazon Kinesis Analytics can assume to write to the destination function on your behalf. You need to grant the necessary permissions to this role.

MappingParameters

mappingParameters_jSONMappingParameters :: Lens' MappingParameters (Maybe JSONMappingParameters)

Provides additional mapping information when JSON is the record format on the streaming source.

mappingParameters_cSVMappingParameters :: Lens' MappingParameters (Maybe CSVMappingParameters)

Provides additional mapping information when the record format uses delimiters (for example, CSV).

Output

output_lambdaOutput :: Lens' Output (Maybe LambdaOutput)

Identifies an AWS Lambda function as the destination.

output_kinesisFirehoseOutput :: Lens' Output (Maybe KinesisFirehoseOutput)

Identifies an Amazon Kinesis Firehose delivery stream as the destination.

output_kinesisStreamsOutput :: Lens' Output (Maybe KinesisStreamsOutput)

Identifies an Amazon Kinesis stream as the destination.

output_name :: Lens' Output Text

Name of the in-application stream.

output_destinationSchema :: Lens' Output DestinationSchema

Describes the data format when records are written to the destination. For more information, see Configuring Application Output.

OutputDescription

outputDescription_kinesisStreamsOutputDescription :: Lens' OutputDescription (Maybe KinesisStreamsOutputDescription)

Describes Amazon Kinesis stream configured as the destination where output is written.

outputDescription_kinesisFirehoseOutputDescription :: Lens' OutputDescription (Maybe KinesisFirehoseOutputDescription)

Describes the Amazon Kinesis Firehose delivery stream configured as the destination where output is written.

outputDescription_destinationSchema :: Lens' OutputDescription (Maybe DestinationSchema)

Data format used for writing data to the destination.

outputDescription_outputId :: Lens' OutputDescription (Maybe Text)

A unique identifier for the output configuration.

outputDescription_name :: Lens' OutputDescription (Maybe Text)

Name of the in-application stream configured as output.

outputDescription_lambdaOutputDescription :: Lens' OutputDescription (Maybe LambdaOutputDescription)

Describes the AWS Lambda function configured as the destination where output is written.

OutputUpdate

outputUpdate_kinesisFirehoseOutputUpdate :: Lens' OutputUpdate (Maybe KinesisFirehoseOutputUpdate)

Describes an Amazon Kinesis Firehose delivery stream as the destination for the output.

outputUpdate_destinationSchemaUpdate :: Lens' OutputUpdate (Maybe DestinationSchema)

Describes the data format when records are written to the destination. For more information, see Configuring Application Output.

outputUpdate_kinesisStreamsOutputUpdate :: Lens' OutputUpdate (Maybe KinesisStreamsOutputUpdate)

Describes an Amazon Kinesis stream as the destination for the output.

outputUpdate_nameUpdate :: Lens' OutputUpdate (Maybe Text)

If you want to specify a different in-application stream for this output configuration, use this field to specify the new in-application stream name.

outputUpdate_lambdaOutputUpdate :: Lens' OutputUpdate (Maybe LambdaOutputUpdate)

Describes an AWS Lambda function as the destination for the output.

outputUpdate_outputId :: Lens' OutputUpdate Text

Identifies the specific output configuration that you want to update.

RecordColumn

recordColumn_mapping :: Lens' RecordColumn (Maybe Text)

Reference to the data element in the streaming input or the reference data source. This element is required if the RecordFormatType is JSON.

recordColumn_name :: Lens' RecordColumn Text

Name of the column created in the in-application input stream or reference table.

recordColumn_sqlType :: Lens' RecordColumn Text

Type of column created in the in-application input stream or reference table.

RecordFormat

recordFormat_mappingParameters :: Lens' RecordFormat (Maybe MappingParameters)

When configuring application input at the time of creating or updating an application, provides additional mapping information specific to the record format (such as JSON, CSV, or record fields delimited by some delimiter) on the streaming source.

recordFormat_recordFormatType :: Lens' RecordFormat RecordFormatType

The type of record format.

ReferenceDataSource

referenceDataSource_s3ReferenceDataSource :: Lens' ReferenceDataSource (Maybe S3ReferenceDataSource)

Identifies the S3 bucket and object that contains the reference data. Also identifies the IAM role Amazon Kinesis Analytics can assume to read this object on your behalf. An Amazon Kinesis Analytics application loads reference data only once. If the data changes, you call the UpdateApplication operation to trigger reloading of data into your application.

referenceDataSource_tableName :: Lens' ReferenceDataSource Text

Name of the in-application table to create.

referenceDataSource_referenceSchema :: Lens' ReferenceDataSource SourceSchema

Describes the format of the data in the streaming source, and how each data element maps to corresponding columns created in the in-application stream.

ReferenceDataSourceDescription

referenceDataSourceDescription_referenceSchema :: Lens' ReferenceDataSourceDescription (Maybe SourceSchema)

Describes the format of the data in the streaming source, and how each data element maps to corresponding columns created in the in-application stream.

referenceDataSourceDescription_referenceId :: Lens' ReferenceDataSourceDescription Text

ID of the reference data source. This is the ID that Amazon Kinesis Analytics assigns when you add the reference data source to your application using the AddApplicationReferenceDataSource operation.

referenceDataSourceDescription_tableName :: Lens' ReferenceDataSourceDescription Text

The in-application table name created by the specific reference data source configuration.

referenceDataSourceDescription_s3ReferenceDataSourceDescription :: Lens' ReferenceDataSourceDescription S3ReferenceDataSourceDescription

Provides the S3 bucket name, the object key name that contains the reference data. It also provides the Amazon Resource Name (ARN) of the IAM role that Amazon Kinesis Analytics can assume to read the Amazon S3 object and populate the in-application reference table.

ReferenceDataSourceUpdate

referenceDataSourceUpdate_s3ReferenceDataSourceUpdate :: Lens' ReferenceDataSourceUpdate (Maybe S3ReferenceDataSourceUpdate)

Describes the S3 bucket name, object key name, and IAM role that Amazon Kinesis Analytics can assume to read the Amazon S3 object on your behalf and populate the in-application reference table.

referenceDataSourceUpdate_referenceSchemaUpdate :: Lens' ReferenceDataSourceUpdate (Maybe SourceSchema)

Describes the format of the data in the streaming source, and how each data element maps to corresponding columns created in the in-application stream.

referenceDataSourceUpdate_tableNameUpdate :: Lens' ReferenceDataSourceUpdate (Maybe Text)

In-application table name that is created by this update.

referenceDataSourceUpdate_referenceId :: Lens' ReferenceDataSourceUpdate Text

ID of the reference data source being updated. You can use the DescribeApplication operation to get this value.

S3Configuration

s3Configuration_roleARN :: Lens' S3Configuration Text

IAM ARN of the role used to access the data.

s3Configuration_bucketARN :: Lens' S3Configuration Text

ARN of the S3 bucket that contains the data.

s3Configuration_fileKey :: Lens' S3Configuration Text

The name of the object that contains the data.

S3ReferenceDataSource

s3ReferenceDataSource_bucketARN :: Lens' S3ReferenceDataSource Text

Amazon Resource Name (ARN) of the S3 bucket.

s3ReferenceDataSource_fileKey :: Lens' S3ReferenceDataSource Text

Object key name containing reference data.

s3ReferenceDataSource_referenceRoleARN :: Lens' S3ReferenceDataSource Text

ARN of the IAM role that the service can assume to read data on your behalf. This role must have permission for the s3:GetObject action on the object and trust policy that allows Amazon Kinesis Analytics service principal to assume this role.

S3ReferenceDataSourceDescription

s3ReferenceDataSourceDescription_bucketARN :: Lens' S3ReferenceDataSourceDescription Text

Amazon Resource Name (ARN) of the S3 bucket.

s3ReferenceDataSourceDescription_fileKey :: Lens' S3ReferenceDataSourceDescription Text

Amazon S3 object key name.

s3ReferenceDataSourceDescription_referenceRoleARN :: Lens' S3ReferenceDataSourceDescription Text

ARN of the IAM role that Amazon Kinesis Analytics can assume to read the Amazon S3 object on your behalf to populate the in-application reference table.

S3ReferenceDataSourceUpdate

s3ReferenceDataSourceUpdate_referenceRoleARNUpdate :: Lens' S3ReferenceDataSourceUpdate (Maybe Text)

ARN of the IAM role that Amazon Kinesis Analytics can assume to read the Amazon S3 object and populate the in-application.

s3ReferenceDataSourceUpdate_fileKeyUpdate :: Lens' S3ReferenceDataSourceUpdate (Maybe Text)

Object key name.

s3ReferenceDataSourceUpdate_bucketARNUpdate :: Lens' S3ReferenceDataSourceUpdate (Maybe Text)

Amazon Resource Name (ARN) of the S3 bucket.

SourceSchema

sourceSchema_recordEncoding :: Lens' SourceSchema (Maybe Text)

Specifies the encoding of the records in the streaming source. For example, UTF-8.

sourceSchema_recordFormat :: Lens' SourceSchema RecordFormat

Specifies the format of the records on the streaming source.

sourceSchema_recordColumns :: Lens' SourceSchema (NonEmpty RecordColumn)

A list of RecordColumn objects.

Tag

tag_value :: Lens' Tag (Maybe Text)

The value of the key-value tag. The value is optional.

tag_key :: Lens' Tag Text

The key of the key-value tag.