Network.AWS.KinesisAnalytics.ListApplications

Description

This documentation is for version 1 of the Amazon Kinesis Data Analytics API, which only supports SQL applications. Version 2 of the API supports SQL and Java applications. For more information about version 2, see Amazon Kinesis Data Analytics API V2 Documentation.

Returns a list of Amazon Kinesis Analytics applications in your account. For each application, the response includes the application name, Amazon Resource Name (ARN), and status. If the response returns the HasMoreApplications value as true, you can send another request by adding the ExclusiveStartApplicationName in the request body, and set the value of this to the last application name from the previous response.

If you want detailed information about a specific application, use DescribeApplication.

This operation requires permissions to perform the kinesisanalytics:ListApplications action.

Creating a Request

data ListApplications

See: newListApplications smart constructor.

Constructors

ListApplications'

Fields

exclusiveStartApplicationName :: Maybe Text

Name of the application to start the list with. When using pagination to retrieve the list, you don't need to specify this parameter in the first request. However, in subsequent requests, you add the last application name from the previous response to get the next page of applications.

limit :: Maybe Natural

Maximum number of applications to list.

newListApplications :: ListApplications

Create a value of ListApplications with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:exclusiveStartApplicationName:ListApplications', listApplications_exclusiveStartApplicationName - Name of the application to start the list with. When using pagination to retrieve the list, you don't need to specify this parameter in the first request. However, in subsequent requests, you add the last application name from the previous response to get the next page of applications.

$sel:limit:ListApplications', listApplications_limit - Maximum number of applications to list.

Request Lenses

listApplications_exclusiveStartApplicationName :: Lens' ListApplications (Maybe Text)

Name of the application to start the list with. When using pagination to retrieve the list, you don't need to specify this parameter in the first request. However, in subsequent requests, you add the last application name from the previous response to get the next page of applications.

listApplications_limit :: Lens' ListApplications (Maybe Natural)

Maximum number of applications to list.

Destructuring the Response

data ListApplicationsResponse

See: newListApplicationsResponse smart constructor.

Constructors

ListApplicationsResponse'

Fields

httpStatus :: Int

The response's http status code.

applicationSummaries :: [ApplicationSummary]

List of ApplicationSummary objects.

hasMoreApplications :: Bool

Returns true if there are more applications to retrieve.

newListApplicationsResponse

Arguments

:: Int
$sel:httpStatus:ListApplicationsResponse'

-> Bool
$sel:hasMoreApplications:ListApplicationsResponse'

-> ListApplicationsResponse

Create a value of ListApplicationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ListApplicationsResponse', listApplicationsResponse_httpStatus - The response's http status code.

$sel:applicationSummaries:ListApplicationsResponse', listApplicationsResponse_applicationSummaries - List of ApplicationSummary objects.

$sel:hasMoreApplications:ListApplicationsResponse', listApplicationsResponse_hasMoreApplications - Returns true if there are more applications to retrieve.

Response Lenses

listApplicationsResponse_httpStatus :: Lens' ListApplicationsResponse Int

The response's http status code.

listApplicationsResponse_applicationSummaries :: Lens' ListApplicationsResponse [ApplicationSummary]

List of ApplicationSummary objects.

listApplicationsResponse_hasMoreApplications :: Lens' ListApplicationsResponse Bool

Returns true if there are more applications to retrieve.