Network.AWS.KinesisAnalytics.Types.ApplicationDetail

Description

Documentation

data ApplicationDetail

This documentation is for version 1 of the Amazon Kinesis Data Analytics API, which only supports SQL applications. Version 2 of the API supports SQL and Java applications. For more information about version 2, see Amazon Kinesis Data Analytics API V2 Documentation.

Provides a description of the application, including the application Amazon Resource Name (ARN), status, latest version, and input and output configuration.

See: newApplicationDetail smart constructor.

Constructors

ApplicationDetail'

Fields

applicationDescription :: Maybe Text

Description of the application.

outputDescriptions :: Maybe [OutputDescription]

Describes the application output configuration. For more information, see Configuring Application Output.

cloudWatchLoggingOptionDescriptions :: Maybe [CloudWatchLoggingOptionDescription]

Describes the CloudWatch log streams that are configured to receive application messages. For more information about using CloudWatch log streams with Amazon Kinesis Analytics applications, see Working with Amazon CloudWatch Logs.

referenceDataSourceDescriptions :: Maybe [ReferenceDataSourceDescription]

Describes reference data sources configured for the application. For more information, see Configuring Application Input.

inputDescriptions :: Maybe [InputDescription]

Describes the application input configuration. For more information, see Configuring Application Input.

applicationCode :: Maybe Text

Returns the application code that you provided to perform data analysis on any of the in-application streams in your application.

createTimestamp :: Maybe POSIX

Time stamp when the application version was created.

lastUpdateTimestamp :: Maybe POSIX

Time stamp when the application was last updated.

applicationName :: Text

Name of the application.

applicationARN :: Text

ARN of the application.

applicationStatus :: ApplicationStatus

Status of the application.

applicationVersionId :: Natural

Provides the current application version.

newApplicationDetail

Arguments

:: Text
$sel:applicationName:ApplicationDetail'

-> Text
$sel:applicationARN:ApplicationDetail'

-> ApplicationStatus
$sel:applicationStatus:ApplicationDetail'

-> Natural
$sel:applicationVersionId:ApplicationDetail'

-> ApplicationDetail

Create a value of ApplicationDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationDescription:ApplicationDetail', applicationDetail_applicationDescription - Description of the application.

$sel:outputDescriptions:ApplicationDetail', applicationDetail_outputDescriptions - Describes the application output configuration. For more information, see Configuring Application Output.

$sel:cloudWatchLoggingOptionDescriptions:ApplicationDetail', applicationDetail_cloudWatchLoggingOptionDescriptions - Describes the CloudWatch log streams that are configured to receive application messages. For more information about using CloudWatch log streams with Amazon Kinesis Analytics applications, see Working with Amazon CloudWatch Logs.

$sel:referenceDataSourceDescriptions:ApplicationDetail', applicationDetail_referenceDataSourceDescriptions - Describes reference data sources configured for the application. For more information, see Configuring Application Input.

$sel:inputDescriptions:ApplicationDetail', applicationDetail_inputDescriptions - Describes the application input configuration. For more information, see Configuring Application Input.

$sel:applicationCode:ApplicationDetail', applicationDetail_applicationCode - Returns the application code that you provided to perform data analysis on any of the in-application streams in your application.

$sel:createTimestamp:ApplicationDetail', applicationDetail_createTimestamp - Time stamp when the application version was created.

$sel:lastUpdateTimestamp:ApplicationDetail', applicationDetail_lastUpdateTimestamp - Time stamp when the application was last updated.

$sel:applicationName:ApplicationDetail', applicationDetail_applicationName - Name of the application.

$sel:applicationARN:ApplicationDetail', applicationDetail_applicationARN - ARN of the application.

$sel:applicationStatus:ApplicationDetail', applicationDetail_applicationStatus - Status of the application.

$sel:applicationVersionId:ApplicationDetail', applicationDetail_applicationVersionId - Provides the current application version.

applicationDetail_applicationDescription :: Lens' ApplicationDetail (Maybe Text)

Description of the application.

applicationDetail_outputDescriptions :: Lens' ApplicationDetail (Maybe [OutputDescription])

Describes the application output configuration. For more information, see Configuring Application Output.

applicationDetail_cloudWatchLoggingOptionDescriptions :: Lens' ApplicationDetail (Maybe [CloudWatchLoggingOptionDescription])

Describes the CloudWatch log streams that are configured to receive application messages. For more information about using CloudWatch log streams with Amazon Kinesis Analytics applications, see Working with Amazon CloudWatch Logs.

applicationDetail_referenceDataSourceDescriptions :: Lens' ApplicationDetail (Maybe [ReferenceDataSourceDescription])

Describes reference data sources configured for the application. For more information, see Configuring Application Input.

applicationDetail_inputDescriptions :: Lens' ApplicationDetail (Maybe [InputDescription])

Describes the application input configuration. For more information, see Configuring Application Input.

applicationDetail_applicationCode :: Lens' ApplicationDetail (Maybe Text)

Returns the application code that you provided to perform data analysis on any of the in-application streams in your application.

applicationDetail_createTimestamp :: Lens' ApplicationDetail (Maybe UTCTime)

Time stamp when the application version was created.

applicationDetail_lastUpdateTimestamp :: Lens' ApplicationDetail (Maybe UTCTime)

Time stamp when the application was last updated.

applicationDetail_applicationName :: Lens' ApplicationDetail Text

Name of the application.

applicationDetail_applicationARN :: Lens' ApplicationDetail Text

ARN of the application.

applicationDetail_applicationStatus :: Lens' ApplicationDetail ApplicationStatus

Status of the application.

applicationDetail_applicationVersionId :: Lens' ApplicationDetail Natural

Provides the current application version.