Network.AWS.KinesisAnalytics.Types.ApplicationUpdate

Description

Documentation

data ApplicationUpdate

Describes updates to apply to an existing Amazon Kinesis Analytics application.

See: newApplicationUpdate smart constructor.

Constructors

ApplicationUpdate'

Fields

referenceDataSourceUpdates :: Maybe [ReferenceDataSourceUpdate]

Describes application reference data source updates.

inputUpdates :: Maybe [InputUpdate]

Describes application input configuration updates.

cloudWatchLoggingOptionUpdates :: Maybe [CloudWatchLoggingOptionUpdate]

Describes application CloudWatch logging option updates.

outputUpdates :: Maybe [OutputUpdate]

Describes application output configuration updates.

applicationCodeUpdate :: Maybe Text

Describes application code updates.

newApplicationUpdate :: ApplicationUpdate

Create a value of ApplicationUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:referenceDataSourceUpdates:ApplicationUpdate', applicationUpdate_referenceDataSourceUpdates - Describes application reference data source updates.

$sel:inputUpdates:ApplicationUpdate', applicationUpdate_inputUpdates - Describes application input configuration updates.

$sel:cloudWatchLoggingOptionUpdates:ApplicationUpdate', applicationUpdate_cloudWatchLoggingOptionUpdates - Describes application CloudWatch logging option updates.

$sel:outputUpdates:ApplicationUpdate', applicationUpdate_outputUpdates - Describes application output configuration updates.

$sel:applicationCodeUpdate:ApplicationUpdate', applicationUpdate_applicationCodeUpdate - Describes application code updates.

applicationUpdate_referenceDataSourceUpdates :: Lens' ApplicationUpdate (Maybe [ReferenceDataSourceUpdate])

Describes application reference data source updates.

applicationUpdate_inputUpdates :: Lens' ApplicationUpdate (Maybe [InputUpdate])

Describes application input configuration updates.

applicationUpdate_cloudWatchLoggingOptionUpdates :: Lens' ApplicationUpdate (Maybe [CloudWatchLoggingOptionUpdate])

Describes application CloudWatch logging option updates.

applicationUpdate_outputUpdates :: Lens' ApplicationUpdate (Maybe [OutputUpdate])

Describes application output configuration updates.

applicationUpdate_applicationCodeUpdate :: Lens' ApplicationUpdate (Maybe Text)

Describes application code updates.