Network.AWS.KinesisAnalytics.Types.InputSchemaUpdate

Description

Documentation

data InputSchemaUpdate

Describes updates for the application's input schema.

See: newInputSchemaUpdate smart constructor.

Constructors

InputSchemaUpdate'

Fields

recordFormatUpdate :: Maybe RecordFormat

Specifies the format of the records on the streaming source.

recordEncodingUpdate :: Maybe Text

Specifies the encoding of the records in the streaming source. For example, UTF-8.

recordColumnUpdates :: Maybe (NonEmpty RecordColumn)

A list of RecordColumn objects. Each object describes the mapping of the streaming source element to the corresponding column in the in-application stream.

newInputSchemaUpdate :: InputSchemaUpdate

Create a value of InputSchemaUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recordFormatUpdate:InputSchemaUpdate', inputSchemaUpdate_recordFormatUpdate - Specifies the format of the records on the streaming source.

$sel:recordEncodingUpdate:InputSchemaUpdate', inputSchemaUpdate_recordEncodingUpdate - Specifies the encoding of the records in the streaming source. For example, UTF-8.

$sel:recordColumnUpdates:InputSchemaUpdate', inputSchemaUpdate_recordColumnUpdates - A list of RecordColumn objects. Each object describes the mapping of the streaming source element to the corresponding column in the in-application stream.

inputSchemaUpdate_recordFormatUpdate :: Lens' InputSchemaUpdate (Maybe RecordFormat)

Specifies the format of the records on the streaming source.

inputSchemaUpdate_recordEncodingUpdate :: Lens' InputSchemaUpdate (Maybe Text)

Specifies the encoding of the records in the streaming source. For example, UTF-8.

inputSchemaUpdate_recordColumnUpdates :: Lens' InputSchemaUpdate (Maybe (NonEmpty RecordColumn))

A list of RecordColumn objects. Each object describes the mapping of the streaming source element to the corresponding column in the in-application stream.