Network.AWS.KinesisAnalytics.Types.KinesisFirehoseOutput

Description

Documentation

data KinesisFirehoseOutput

When configuring application output, identifies an Amazon Kinesis Firehose delivery stream as the destination. You provide the stream Amazon Resource Name (ARN) and an IAM role that enables Amazon Kinesis Analytics to write to the stream on your behalf.

See: newKinesisFirehoseOutput smart constructor.

Constructors

KinesisFirehoseOutput'

Fields

resourceARN :: Text

ARN of the destination Amazon Kinesis Firehose delivery stream to write to.

roleARN :: Text

ARN of the IAM role that Amazon Kinesis Analytics can assume to write to the destination stream on your behalf. You need to grant the necessary permissions to this role.

newKinesisFirehoseOutput

Arguments

:: Text
$sel:resourceARN:KinesisFirehoseOutput'

-> Text
$sel:roleARN:KinesisFirehoseOutput'

-> KinesisFirehoseOutput

Create a value of KinesisFirehoseOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceARN:KinesisFirehoseOutput', kinesisFirehoseOutput_resourceARN - ARN of the destination Amazon Kinesis Firehose delivery stream to write to.

$sel:roleARN:KinesisFirehoseOutput', kinesisFirehoseOutput_roleARN - ARN of the IAM role that Amazon Kinesis Analytics can assume to write to the destination stream on your behalf. You need to grant the necessary permissions to this role.

kinesisFirehoseOutput_resourceARN :: Lens' KinesisFirehoseOutput Text

ARN of the destination Amazon Kinesis Firehose delivery stream to write to.

kinesisFirehoseOutput_roleARN :: Lens' KinesisFirehoseOutput Text

ARN of the IAM role that Amazon Kinesis Analytics can assume to write to the destination stream on your behalf. You need to grant the necessary permissions to this role.