Network.AWS.KinesisAnalytics.Types.ReferenceDataSource

Description

Documentation

data ReferenceDataSource

Describes the reference data source by providing the source information (S3 bucket name and object key name), the resulting in-application table name that is created, and the necessary schema to map the data elements in the Amazon S3 object to the in-application table.

See: newReferenceDataSource smart constructor.

Constructors

ReferenceDataSource'

Fields

s3ReferenceDataSource :: Maybe S3ReferenceDataSource

Identifies the S3 bucket and object that contains the reference data. Also identifies the IAM role Amazon Kinesis Analytics can assume to read this object on your behalf. An Amazon Kinesis Analytics application loads reference data only once. If the data changes, you call the UpdateApplication operation to trigger reloading of data into your application.

tableName :: Text

Name of the in-application table to create.

referenceSchema :: SourceSchema

Describes the format of the data in the streaming source, and how each data element maps to corresponding columns created in the in-application stream.

newReferenceDataSource

Arguments

:: Text
$sel:tableName:ReferenceDataSource'

-> SourceSchema
$sel:referenceSchema:ReferenceDataSource'

-> ReferenceDataSource

Create a value of ReferenceDataSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3ReferenceDataSource:ReferenceDataSource', referenceDataSource_s3ReferenceDataSource - Identifies the S3 bucket and object that contains the reference data. Also identifies the IAM role Amazon Kinesis Analytics can assume to read this object on your behalf. An Amazon Kinesis Analytics application loads reference data only once. If the data changes, you call the UpdateApplication operation to trigger reloading of data into your application.

$sel:tableName:ReferenceDataSource', referenceDataSource_tableName - Name of the in-application table to create.

$sel:referenceSchema:ReferenceDataSource', referenceDataSource_referenceSchema - Describes the format of the data in the streaming source, and how each data element maps to corresponding columns created in the in-application stream.

referenceDataSource_s3ReferenceDataSource :: Lens' ReferenceDataSource (Maybe S3ReferenceDataSource)

Identifies the S3 bucket and object that contains the reference data. Also identifies the IAM role Amazon Kinesis Analytics can assume to read this object on your behalf. An Amazon Kinesis Analytics application loads reference data only once. If the data changes, you call the UpdateApplication operation to trigger reloading of data into your application.

referenceDataSource_tableName :: Lens' ReferenceDataSource Text

Name of the in-application table to create.

referenceDataSource_referenceSchema :: Lens' ReferenceDataSource SourceSchema

Describes the format of the data in the streaming source, and how each data element maps to corresponding columns created in the in-application stream.