Network.AWS.KinesisAnalytics.Types.ReferenceDataSourceDescription

Description

Documentation

data ReferenceDataSourceDescription

Describes the reference data source configured for an application.

See: newReferenceDataSourceDescription smart constructor.

Constructors

ReferenceDataSourceDescription'

Fields

referenceSchema :: Maybe SourceSchema

Describes the format of the data in the streaming source, and how each data element maps to corresponding columns created in the in-application stream.

referenceId :: Text

ID of the reference data source. This is the ID that Amazon Kinesis Analytics assigns when you add the reference data source to your application using the AddApplicationReferenceDataSource operation.

tableName :: Text

The in-application table name created by the specific reference data source configuration.

s3ReferenceDataSourceDescription :: S3ReferenceDataSourceDescription

Provides the S3 bucket name, the object key name that contains the reference data. It also provides the Amazon Resource Name (ARN) of the IAM role that Amazon Kinesis Analytics can assume to read the Amazon S3 object and populate the in-application reference table.

newReferenceDataSourceDescription

Arguments

:: Text
$sel:referenceId:ReferenceDataSourceDescription'

-> Text
$sel:tableName:ReferenceDataSourceDescription'

-> S3ReferenceDataSourceDescription
$sel:s3ReferenceDataSourceDescription:ReferenceDataSourceDescription'

-> ReferenceDataSourceDescription

Create a value of ReferenceDataSourceDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:referenceSchema:ReferenceDataSourceDescription', referenceDataSourceDescription_referenceSchema - Describes the format of the data in the streaming source, and how each data element maps to corresponding columns created in the in-application stream.

$sel:referenceId:ReferenceDataSourceDescription', referenceDataSourceDescription_referenceId - ID of the reference data source. This is the ID that Amazon Kinesis Analytics assigns when you add the reference data source to your application using the AddApplicationReferenceDataSource operation.

$sel:tableName:ReferenceDataSourceDescription', referenceDataSourceDescription_tableName - The in-application table name created by the specific reference data source configuration.

$sel:s3ReferenceDataSourceDescription:ReferenceDataSourceDescription', referenceDataSourceDescription_s3ReferenceDataSourceDescription - Provides the S3 bucket name, the object key name that contains the reference data. It also provides the Amazon Resource Name (ARN) of the IAM role that Amazon Kinesis Analytics can assume to read the Amazon S3 object and populate the in-application reference table.

referenceDataSourceDescription_referenceSchema :: Lens' ReferenceDataSourceDescription (Maybe SourceSchema)

Describes the format of the data in the streaming source, and how each data element maps to corresponding columns created in the in-application stream.

referenceDataSourceDescription_referenceId :: Lens' ReferenceDataSourceDescription Text

ID of the reference data source. This is the ID that Amazon Kinesis Analytics assigns when you add the reference data source to your application using the AddApplicationReferenceDataSource operation.

referenceDataSourceDescription_tableName :: Lens' ReferenceDataSourceDescription Text

The in-application table name created by the specific reference data source configuration.

referenceDataSourceDescription_s3ReferenceDataSourceDescription :: Lens' ReferenceDataSourceDescription S3ReferenceDataSourceDescription

Provides the S3 bucket name, the object key name that contains the reference data. It also provides the Amazon Resource Name (ARN) of the IAM role that Amazon Kinesis Analytics can assume to read the Amazon S3 object and populate the in-application reference table.