Network.AWS.KinesisAnalyticsV2.AddApplicationInput

Description

Adds a streaming source to your SQL-based Kinesis Data Analytics application.

You can add a streaming source when you create an application, or you can use this operation to add a streaming source after you create an application. For more information, see CreateApplication.

Any configuration update, including adding a streaming source using this operation, results in a new version of the application. You can use the DescribeApplication operation to find the current application version.

Creating a Request

data AddApplicationInput

See: newAddApplicationInput smart constructor.

Constructors

AddApplicationInput'

Fields

applicationName :: Text

The name of your existing application to which you want to add the streaming source.

currentApplicationVersionId :: Natural

The current version of your application. You must provide the ApplicationVersionID or the ConditionalToken.You can use the DescribeApplication operation to find the current application version.

input :: Input

The Input to add.

newAddApplicationInput

Arguments

:: Text
$sel:applicationName:AddApplicationInput'

-> Natural
$sel:currentApplicationVersionId:AddApplicationInput'

-> Input
$sel:input:AddApplicationInput'

-> AddApplicationInput

Create a value of AddApplicationInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationName:AddApplicationInput', addApplicationInput_applicationName - The name of your existing application to which you want to add the streaming source.

$sel:currentApplicationVersionId:AddApplicationInput', addApplicationInput_currentApplicationVersionId - The current version of your application. You must provide the ApplicationVersionID or the ConditionalToken.You can use the DescribeApplication operation to find the current application version.

$sel:input:AddApplicationInput', addApplicationInput_input - The Input to add.

Request Lenses

addApplicationInput_applicationName :: Lens' AddApplicationInput Text

The name of your existing application to which you want to add the streaming source.

addApplicationInput_currentApplicationVersionId :: Lens' AddApplicationInput Natural

The current version of your application. You must provide the ApplicationVersionID or the ConditionalToken.You can use the DescribeApplication operation to find the current application version.

addApplicationInput_input :: Lens' AddApplicationInput Input

The Input to add.

Destructuring the Response

data AddApplicationInputResponse

See: newAddApplicationInputResponse smart constructor.

Constructors

AddApplicationInputResponse'

Fields

applicationARN :: Maybe Text

The Amazon Resource Name (ARN) of the application.

applicationVersionId :: Maybe Natural

Provides the current application version.

inputDescriptions :: Maybe [InputDescription]

Describes the application input configuration.

httpStatus :: Int

The response's http status code.

newAddApplicationInputResponse

Arguments

:: Int
$sel:httpStatus:AddApplicationInputResponse'

-> AddApplicationInputResponse

Create a value of AddApplicationInputResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationARN:AddApplicationInputResponse', addApplicationInputResponse_applicationARN - The Amazon Resource Name (ARN) of the application.

$sel:applicationVersionId:AddApplicationInputResponse', addApplicationInputResponse_applicationVersionId - Provides the current application version.

$sel:inputDescriptions:AddApplicationInputResponse', addApplicationInputResponse_inputDescriptions - Describes the application input configuration.

$sel:httpStatus:AddApplicationInputResponse', addApplicationInputResponse_httpStatus - The response's http status code.

Response Lenses

addApplicationInputResponse_applicationARN :: Lens' AddApplicationInputResponse (Maybe Text)

The Amazon Resource Name (ARN) of the application.

addApplicationInputResponse_applicationVersionId :: Lens' AddApplicationInputResponse (Maybe Natural)

Provides the current application version.

addApplicationInputResponse_inputDescriptions :: Lens' AddApplicationInputResponse (Maybe [InputDescription])

Describes the application input configuration.

addApplicationInputResponse_httpStatus :: Lens' AddApplicationInputResponse Int

The response's http status code.