Network.AWS.KinesisAnalyticsV2.AddApplicationInputProcessingConfiguration

Description

Adds an InputProcessingConfiguration to a SQL-based Kinesis Data Analytics application. An input processor pre-processes records on the input stream before the application's SQL code executes. Currently, the only input processor available is Amazon Lambda.

Creating a Request

data AddApplicationInputProcessingConfiguration

See: newAddApplicationInputProcessingConfiguration smart constructor.

Constructors

AddApplicationInputProcessingConfiguration'

Fields

applicationName :: Text

The name of the application to which you want to add the input processing configuration.

currentApplicationVersionId :: Natural

The version of the application to which you want to add the input processing configuration. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

inputId :: Text

The ID of the input configuration to add the input processing configuration to. You can get a list of the input IDs for an application using the DescribeApplication operation.

inputProcessingConfiguration :: InputProcessingConfiguration

The InputProcessingConfiguration to add to the application.

newAddApplicationInputProcessingConfiguration

Arguments

:: Text
$sel:applicationName:AddApplicationInputProcessingConfiguration'

-> Natural
$sel:currentApplicationVersionId:AddApplicationInputProcessingConfiguration'

-> Text
$sel:inputId:AddApplicationInputProcessingConfiguration'

-> InputProcessingConfiguration
$sel:inputProcessingConfiguration:AddApplicationInputProcessingConfiguration'

-> AddApplicationInputProcessingConfiguration

Create a value of AddApplicationInputProcessingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationName:AddApplicationInputProcessingConfiguration', addApplicationInputProcessingConfiguration_applicationName - The name of the application to which you want to add the input processing configuration.

$sel:currentApplicationVersionId:AddApplicationInputProcessingConfiguration', addApplicationInputProcessingConfiguration_currentApplicationVersionId - The version of the application to which you want to add the input processing configuration. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

$sel:inputId:AddApplicationInputProcessingConfiguration', addApplicationInputProcessingConfiguration_inputId - The ID of the input configuration to add the input processing configuration to. You can get a list of the input IDs for an application using the DescribeApplication operation.

$sel:inputProcessingConfiguration:AddApplicationInputProcessingConfiguration', addApplicationInputProcessingConfiguration_inputProcessingConfiguration - The InputProcessingConfiguration to add to the application.

Request Lenses

addApplicationInputProcessingConfiguration_applicationName :: Lens' AddApplicationInputProcessingConfiguration Text

The name of the application to which you want to add the input processing configuration.

addApplicationInputProcessingConfiguration_currentApplicationVersionId :: Lens' AddApplicationInputProcessingConfiguration Natural

The version of the application to which you want to add the input processing configuration. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

addApplicationInputProcessingConfiguration_inputId :: Lens' AddApplicationInputProcessingConfiguration Text

The ID of the input configuration to add the input processing configuration to. You can get a list of the input IDs for an application using the DescribeApplication operation.

addApplicationInputProcessingConfiguration_inputProcessingConfiguration :: Lens' AddApplicationInputProcessingConfiguration InputProcessingConfiguration

The InputProcessingConfiguration to add to the application.

Destructuring the Response

data AddApplicationInputProcessingConfigurationResponse

See: newAddApplicationInputProcessingConfigurationResponse smart constructor.

Constructors

AddApplicationInputProcessingConfigurationResponse'

Fields

applicationARN :: Maybe Text

The Amazon Resource Name (ARN) of the application.

applicationVersionId :: Maybe Natural

Provides the current application version.

inputId :: Maybe Text

The input ID that is associated with the application input. This is the ID that Kinesis Data Analytics assigns to each input configuration that you add to your application.

inputProcessingConfigurationDescription :: Maybe InputProcessingConfigurationDescription

The description of the preprocessor that executes on records in this input before the application's code is run.

httpStatus :: Int

The response's http status code.

newAddApplicationInputProcessingConfigurationResponse

Arguments

:: Int
$sel:httpStatus:AddApplicationInputProcessingConfigurationResponse'

-> AddApplicationInputProcessingConfigurationResponse

Create a value of AddApplicationInputProcessingConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationARN:AddApplicationInputProcessingConfigurationResponse', addApplicationInputProcessingConfigurationResponse_applicationARN - The Amazon Resource Name (ARN) of the application.

$sel:applicationVersionId:AddApplicationInputProcessingConfigurationResponse', addApplicationInputProcessingConfigurationResponse_applicationVersionId - Provides the current application version.

$sel:inputId:AddApplicationInputProcessingConfiguration', addApplicationInputProcessingConfigurationResponse_inputId - The input ID that is associated with the application input. This is the ID that Kinesis Data Analytics assigns to each input configuration that you add to your application.

$sel:inputProcessingConfigurationDescription:AddApplicationInputProcessingConfigurationResponse', addApplicationInputProcessingConfigurationResponse_inputProcessingConfigurationDescription - The description of the preprocessor that executes on records in this input before the application's code is run.

$sel:httpStatus:AddApplicationInputProcessingConfigurationResponse', addApplicationInputProcessingConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

addApplicationInputProcessingConfigurationResponse_applicationARN :: Lens' AddApplicationInputProcessingConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the application.

addApplicationInputProcessingConfigurationResponse_applicationVersionId :: Lens' AddApplicationInputProcessingConfigurationResponse (Maybe Natural)

Provides the current application version.

addApplicationInputProcessingConfigurationResponse_inputId :: Lens' AddApplicationInputProcessingConfigurationResponse (Maybe Text)

The input ID that is associated with the application input. This is the ID that Kinesis Data Analytics assigns to each input configuration that you add to your application.

addApplicationInputProcessingConfigurationResponse_inputProcessingConfigurationDescription :: Lens' AddApplicationInputProcessingConfigurationResponse (Maybe InputProcessingConfigurationDescription)

The description of the preprocessor that executes on records in this input before the application's code is run.

addApplicationInputProcessingConfigurationResponse_httpStatus :: Lens' AddApplicationInputProcessingConfigurationResponse Int

The response's http status code.