Network.AWS.KinesisAnalyticsV2.CreateApplicationPresignedUrl

Description

Creates and returns a URL that you can use to connect to an application's extension. Currently, the only available extension is the Apache Flink dashboard.

The IAM role or user used to call this API defines the permissions to access the extension. After the presigned URL is created, no additional permission is required to access this URL. IAM authorization policies for this API are also enforced for every HTTP request that attempts to connect to the extension.

You control the amount of time that the URL will be valid using the SessionExpirationDurationInSeconds parameter. If you do not provide this parameter, the returned URL is valid for twelve hours.

The URL that you get from a call to CreateApplicationPresignedUrl must be used within 3 minutes to be valid. If you first try to use the URL after the 3-minute limit expires, the service returns an HTTP 403 Forbidden error.

Creating a Request

data CreateApplicationPresignedUrl

See: newCreateApplicationPresignedUrl smart constructor.

Constructors

CreateApplicationPresignedUrl'

Fields

sessionExpirationDurationInSeconds :: Maybe Natural

The duration in seconds for which the returned URL will be valid.

applicationName :: Text

The name of the application.

urlType :: UrlType

The type of the extension for which to create and return a URL. Currently, the only valid extension URL type is FLINK_DASHBOARD_URL.

newCreateApplicationPresignedUrl

Arguments

:: Text
$sel:applicationName:CreateApplicationPresignedUrl'

-> UrlType
$sel:urlType:CreateApplicationPresignedUrl'

-> CreateApplicationPresignedUrl

Create a value of CreateApplicationPresignedUrl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sessionExpirationDurationInSeconds:CreateApplicationPresignedUrl', createApplicationPresignedUrl_sessionExpirationDurationInSeconds - The duration in seconds for which the returned URL will be valid.

$sel:applicationName:CreateApplicationPresignedUrl', createApplicationPresignedUrl_applicationName - The name of the application.

$sel:urlType:CreateApplicationPresignedUrl', createApplicationPresignedUrl_urlType - The type of the extension for which to create and return a URL. Currently, the only valid extension URL type is FLINK_DASHBOARD_URL.

Request Lenses

createApplicationPresignedUrl_sessionExpirationDurationInSeconds :: Lens' CreateApplicationPresignedUrl (Maybe Natural)

The duration in seconds for which the returned URL will be valid.

createApplicationPresignedUrl_applicationName :: Lens' CreateApplicationPresignedUrl Text

The name of the application.

createApplicationPresignedUrl_urlType :: Lens' CreateApplicationPresignedUrl UrlType

The type of the extension for which to create and return a URL. Currently, the only valid extension URL type is FLINK_DASHBOARD_URL.

Destructuring the Response

data CreateApplicationPresignedUrlResponse

See: newCreateApplicationPresignedUrlResponse smart constructor.

Constructors

CreateApplicationPresignedUrlResponse'

Fields

authorizedUrl :: Maybe Text

The URL of the extension.

httpStatus :: Int

The response's http status code.

newCreateApplicationPresignedUrlResponse

Arguments

:: Int
$sel:httpStatus:CreateApplicationPresignedUrlResponse'

-> CreateApplicationPresignedUrlResponse

Create a value of CreateApplicationPresignedUrlResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authorizedUrl:CreateApplicationPresignedUrlResponse', createApplicationPresignedUrlResponse_authorizedUrl - The URL of the extension.

$sel:httpStatus:CreateApplicationPresignedUrlResponse', createApplicationPresignedUrlResponse_httpStatus - The response's http status code.

Response Lenses

createApplicationPresignedUrlResponse_authorizedUrl :: Lens' CreateApplicationPresignedUrlResponse (Maybe Text)

The URL of the extension.

createApplicationPresignedUrlResponse_httpStatus :: Lens' CreateApplicationPresignedUrlResponse Int

The response's http status code.