Network.AWS.KinesisAnalyticsV2.DeleteApplicationSnapshot

Description

Deletes a snapshot of application state.

Creating a Request

data DeleteApplicationSnapshot

See: newDeleteApplicationSnapshot smart constructor.

Constructors

DeleteApplicationSnapshot'

Fields

applicationName :: Text

The name of an existing application.

snapshotName :: Text

The identifier for the snapshot delete.

snapshotCreationTimestamp :: POSIX

The creation timestamp of the application snapshot to delete. You can retrieve this value using or .

newDeleteApplicationSnapshot

Arguments

:: Text
$sel:applicationName:DeleteApplicationSnapshot'

-> Text
$sel:snapshotName:DeleteApplicationSnapshot'

-> UTCTime
$sel:snapshotCreationTimestamp:DeleteApplicationSnapshot'

-> DeleteApplicationSnapshot

Create a value of DeleteApplicationSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationName:DeleteApplicationSnapshot', deleteApplicationSnapshot_applicationName - The name of an existing application.

$sel:snapshotName:DeleteApplicationSnapshot', deleteApplicationSnapshot_snapshotName - The identifier for the snapshot delete.

$sel:snapshotCreationTimestamp:DeleteApplicationSnapshot', deleteApplicationSnapshot_snapshotCreationTimestamp - The creation timestamp of the application snapshot to delete. You can retrieve this value using or .

Request Lenses

deleteApplicationSnapshot_applicationName :: Lens' DeleteApplicationSnapshot Text

The name of an existing application.

deleteApplicationSnapshot_snapshotName :: Lens' DeleteApplicationSnapshot Text

The identifier for the snapshot delete.

deleteApplicationSnapshot_snapshotCreationTimestamp :: Lens' DeleteApplicationSnapshot UTCTime

The creation timestamp of the application snapshot to delete. You can retrieve this value using or .

Destructuring the Response

data DeleteApplicationSnapshotResponse

See: newDeleteApplicationSnapshotResponse smart constructor.

Constructors

DeleteApplicationSnapshotResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteApplicationSnapshotResponse

Arguments

:: Int
$sel:httpStatus:DeleteApplicationSnapshotResponse'

-> DeleteApplicationSnapshotResponse

Create a value of DeleteApplicationSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteApplicationSnapshotResponse', deleteApplicationSnapshotResponse_httpStatus - The response's http status code.

Response Lenses

deleteApplicationSnapshotResponse_httpStatus :: Lens' DeleteApplicationSnapshotResponse Int

The response's http status code.