Network.AWS.KinesisAnalyticsV2.DescribeApplicationSnapshot

Description

Returns information about a snapshot of application state data.

Creating a Request

data DescribeApplicationSnapshot

See: newDescribeApplicationSnapshot smart constructor.

Constructors

DescribeApplicationSnapshot'

Fields

applicationName :: Text

The name of an existing application.

snapshotName :: Text

The identifier of an application snapshot. You can retrieve this value using .

newDescribeApplicationSnapshot

Arguments

:: Text
$sel:applicationName:DescribeApplicationSnapshot'

-> Text
$sel:snapshotName:DescribeApplicationSnapshot'

-> DescribeApplicationSnapshot

Create a value of DescribeApplicationSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationName:DescribeApplicationSnapshot', describeApplicationSnapshot_applicationName - The name of an existing application.

$sel:snapshotName:DescribeApplicationSnapshot', describeApplicationSnapshot_snapshotName - The identifier of an application snapshot. You can retrieve this value using .

Request Lenses

describeApplicationSnapshot_applicationName :: Lens' DescribeApplicationSnapshot Text

The name of an existing application.

describeApplicationSnapshot_snapshotName :: Lens' DescribeApplicationSnapshot Text

The identifier of an application snapshot. You can retrieve this value using .

Destructuring the Response

data DescribeApplicationSnapshotResponse

See: newDescribeApplicationSnapshotResponse smart constructor.

Constructors

DescribeApplicationSnapshotResponse'

Fields

httpStatus :: Int

The response's http status code.

snapshotDetails :: SnapshotDetails

An object containing information about the application snapshot.

newDescribeApplicationSnapshotResponse

Arguments

:: Int
$sel:httpStatus:DescribeApplicationSnapshotResponse'

-> SnapshotDetails
$sel:snapshotDetails:DescribeApplicationSnapshotResponse'

-> DescribeApplicationSnapshotResponse

Create a value of DescribeApplicationSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeApplicationSnapshotResponse', describeApplicationSnapshotResponse_httpStatus - The response's http status code.

$sel:snapshotDetails:DescribeApplicationSnapshotResponse', describeApplicationSnapshotResponse_snapshotDetails - An object containing information about the application snapshot.

Response Lenses

describeApplicationSnapshotResponse_httpStatus :: Lens' DescribeApplicationSnapshotResponse Int

The response's http status code.

describeApplicationSnapshotResponse_snapshotDetails :: Lens' DescribeApplicationSnapshotResponse SnapshotDetails

An object containing information about the application snapshot.