Network.AWS.KinesisAnalyticsV2.ListApplicationSnapshots

Description

Lists information about the current application snapshots.

This operation returns paginated results.

Creating a Request

data ListApplicationSnapshots

See: newListApplicationSnapshots smart constructor.

Constructors

ListApplicationSnapshots'

Fields

nextToken :: Maybe Text

Use this parameter if you receive a NextToken response in a previous request that indicates that there is more output available. Set it to the value of the previous call's NextToken response to indicate where the output should continue from.

limit :: Maybe Natural

The maximum number of application snapshots to list.

applicationName :: Text

The name of an existing application.

newListApplicationSnapshots

Arguments

:: Text
$sel:applicationName:ListApplicationSnapshots'

-> ListApplicationSnapshots

Create a value of ListApplicationSnapshots with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListApplicationSnapshots', listApplicationSnapshots_nextToken - Use this parameter if you receive a NextToken response in a previous request that indicates that there is more output available. Set it to the value of the previous call's NextToken response to indicate where the output should continue from.

$sel:limit:ListApplicationSnapshots', listApplicationSnapshots_limit - The maximum number of application snapshots to list.

$sel:applicationName:ListApplicationSnapshots', listApplicationSnapshots_applicationName - The name of an existing application.

Request Lenses

listApplicationSnapshots_nextToken :: Lens' ListApplicationSnapshots (Maybe Text)

Use this parameter if you receive a NextToken response in a previous request that indicates that there is more output available. Set it to the value of the previous call's NextToken response to indicate where the output should continue from.

listApplicationSnapshots_limit :: Lens' ListApplicationSnapshots (Maybe Natural)

The maximum number of application snapshots to list.

listApplicationSnapshots_applicationName :: Lens' ListApplicationSnapshots Text

The name of an existing application.

Destructuring the Response

data ListApplicationSnapshotsResponse

See: newListApplicationSnapshotsResponse smart constructor.

Constructors

ListApplicationSnapshotsResponse'

Fields

snapshotSummaries :: Maybe [SnapshotDetails]

A collection of objects containing information about the application snapshots.

nextToken :: Maybe Text

The token for the next set of results, or null if there are no additional results.

httpStatus :: Int

The response's http status code.

newListApplicationSnapshotsResponse

Arguments

:: Int
$sel:httpStatus:ListApplicationSnapshotsResponse'

-> ListApplicationSnapshotsResponse

Create a value of ListApplicationSnapshotsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshotSummaries:ListApplicationSnapshotsResponse', listApplicationSnapshotsResponse_snapshotSummaries - A collection of objects containing information about the application snapshots.

$sel:nextToken:ListApplicationSnapshots', listApplicationSnapshotsResponse_nextToken - The token for the next set of results, or null if there are no additional results.

$sel:httpStatus:ListApplicationSnapshotsResponse', listApplicationSnapshotsResponse_httpStatus - The response's http status code.

Response Lenses

listApplicationSnapshotsResponse_snapshotSummaries :: Lens' ListApplicationSnapshotsResponse (Maybe [SnapshotDetails])

A collection of objects containing information about the application snapshots.

listApplicationSnapshotsResponse_nextToken :: Lens' ListApplicationSnapshotsResponse (Maybe Text)

The token for the next set of results, or null if there are no additional results.

listApplicationSnapshotsResponse_httpStatus :: Lens' ListApplicationSnapshotsResponse Int

The response's http status code.