Network.AWS.KinesisAnalyticsV2.Types.ApplicationConfiguration

Description

Documentation

data ApplicationConfiguration

Specifies the creation parameters for a Kinesis Data Analytics application.

See: newApplicationConfiguration smart constructor.

Constructors

ApplicationConfiguration'

Fields

applicationSnapshotConfiguration :: Maybe ApplicationSnapshotConfiguration

Describes whether snapshots are enabled for a Flink-based Kinesis Data Analytics application.

vpcConfigurations :: Maybe [VpcConfiguration]

The array of descriptions of VPC configurations available to the application.

environmentProperties :: Maybe EnvironmentProperties

Describes execution properties for a Flink-based Kinesis Data Analytics application.

flinkApplicationConfiguration :: Maybe FlinkApplicationConfiguration

The creation and update parameters for a Flink-based Kinesis Data Analytics application.

sqlApplicationConfiguration :: Maybe SqlApplicationConfiguration

The creation and update parameters for a SQL-based Kinesis Data Analytics application.

applicationCodeConfiguration :: Maybe ApplicationCodeConfiguration

The code location and type parameters for a Flink-based Kinesis Data Analytics application.

zeppelinApplicationConfiguration :: Maybe ZeppelinApplicationConfiguration

The configuration parameters for a Kinesis Data Analytics Studio notebook.

newApplicationConfiguration :: ApplicationConfiguration

Create a value of ApplicationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationSnapshotConfiguration:ApplicationConfiguration', applicationConfiguration_applicationSnapshotConfiguration - Describes whether snapshots are enabled for a Flink-based Kinesis Data Analytics application.

$sel:vpcConfigurations:ApplicationConfiguration', applicationConfiguration_vpcConfigurations - The array of descriptions of VPC configurations available to the application.

$sel:environmentProperties:ApplicationConfiguration', applicationConfiguration_environmentProperties - Describes execution properties for a Flink-based Kinesis Data Analytics application.

$sel:flinkApplicationConfiguration:ApplicationConfiguration', applicationConfiguration_flinkApplicationConfiguration - The creation and update parameters for a Flink-based Kinesis Data Analytics application.

$sel:sqlApplicationConfiguration:ApplicationConfiguration', applicationConfiguration_sqlApplicationConfiguration - The creation and update parameters for a SQL-based Kinesis Data Analytics application.

$sel:applicationCodeConfiguration:ApplicationConfiguration', applicationConfiguration_applicationCodeConfiguration - The code location and type parameters for a Flink-based Kinesis Data Analytics application.

$sel:zeppelinApplicationConfiguration:ApplicationConfiguration', applicationConfiguration_zeppelinApplicationConfiguration - The configuration parameters for a Kinesis Data Analytics Studio notebook.

applicationConfiguration_applicationSnapshotConfiguration :: Lens' ApplicationConfiguration (Maybe ApplicationSnapshotConfiguration)

Describes whether snapshots are enabled for a Flink-based Kinesis Data Analytics application.

applicationConfiguration_vpcConfigurations :: Lens' ApplicationConfiguration (Maybe [VpcConfiguration])

The array of descriptions of VPC configurations available to the application.

applicationConfiguration_environmentProperties :: Lens' ApplicationConfiguration (Maybe EnvironmentProperties)

Describes execution properties for a Flink-based Kinesis Data Analytics application.

applicationConfiguration_flinkApplicationConfiguration :: Lens' ApplicationConfiguration (Maybe FlinkApplicationConfiguration)

The creation and update parameters for a Flink-based Kinesis Data Analytics application.

applicationConfiguration_sqlApplicationConfiguration :: Lens' ApplicationConfiguration (Maybe SqlApplicationConfiguration)

The creation and update parameters for a SQL-based Kinesis Data Analytics application.

applicationConfiguration_applicationCodeConfiguration :: Lens' ApplicationConfiguration (Maybe ApplicationCodeConfiguration)

The code location and type parameters for a Flink-based Kinesis Data Analytics application.

applicationConfiguration_zeppelinApplicationConfiguration :: Lens' ApplicationConfiguration (Maybe ZeppelinApplicationConfiguration)

The configuration parameters for a Kinesis Data Analytics Studio notebook.