Network.AWS.KinesisAnalyticsV2.Types.ApplicationConfigurationDescription

Description

Documentation

data ApplicationConfigurationDescription

Describes details about the application code and starting parameters for a Kinesis Data Analytics application.

See: newApplicationConfigurationDescription smart constructor.

Constructors

ApplicationConfigurationDescription'

Fields

zeppelinApplicationConfigurationDescription :: Maybe ZeppelinApplicationConfigurationDescription

The configuration parameters for a Kinesis Data Analytics Studio notebook.

vpcConfigurationDescriptions :: Maybe [VpcConfigurationDescription]

The array of descriptions of VPC configurations available to the application.

applicationCodeConfigurationDescription :: Maybe ApplicationCodeConfigurationDescription

The details about the application code for a Flink-based Kinesis Data Analytics application.

environmentPropertyDescriptions :: Maybe EnvironmentPropertyDescriptions

Describes execution properties for a Flink-based Kinesis Data Analytics application.

runConfigurationDescription :: Maybe RunConfigurationDescription

The details about the starting properties for a Kinesis Data Analytics application.

sqlApplicationConfigurationDescription :: Maybe SqlApplicationConfigurationDescription

The details about inputs, outputs, and reference data sources for a SQL-based Kinesis Data Analytics application.

flinkApplicationConfigurationDescription :: Maybe FlinkApplicationConfigurationDescription

The details about a Flink-based Kinesis Data Analytics application.

applicationSnapshotConfigurationDescription :: Maybe ApplicationSnapshotConfigurationDescription

Describes whether snapshots are enabled for a Flink-based Kinesis Data Analytics application.

newApplicationConfigurationDescription :: ApplicationConfigurationDescription

Create a value of ApplicationConfigurationDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:zeppelinApplicationConfigurationDescription:ApplicationConfigurationDescription', applicationConfigurationDescription_zeppelinApplicationConfigurationDescription - The configuration parameters for a Kinesis Data Analytics Studio notebook.

$sel:vpcConfigurationDescriptions:ApplicationConfigurationDescription', applicationConfigurationDescription_vpcConfigurationDescriptions - The array of descriptions of VPC configurations available to the application.

$sel:applicationCodeConfigurationDescription:ApplicationConfigurationDescription', applicationConfigurationDescription_applicationCodeConfigurationDescription - The details about the application code for a Flink-based Kinesis Data Analytics application.

$sel:environmentPropertyDescriptions:ApplicationConfigurationDescription', applicationConfigurationDescription_environmentPropertyDescriptions - Describes execution properties for a Flink-based Kinesis Data Analytics application.

$sel:runConfigurationDescription:ApplicationConfigurationDescription', applicationConfigurationDescription_runConfigurationDescription - The details about the starting properties for a Kinesis Data Analytics application.

$sel:sqlApplicationConfigurationDescription:ApplicationConfigurationDescription', applicationConfigurationDescription_sqlApplicationConfigurationDescription - The details about inputs, outputs, and reference data sources for a SQL-based Kinesis Data Analytics application.

$sel:flinkApplicationConfigurationDescription:ApplicationConfigurationDescription', applicationConfigurationDescription_flinkApplicationConfigurationDescription - The details about a Flink-based Kinesis Data Analytics application.

$sel:applicationSnapshotConfigurationDescription:ApplicationConfigurationDescription', applicationConfigurationDescription_applicationSnapshotConfigurationDescription - Describes whether snapshots are enabled for a Flink-based Kinesis Data Analytics application.

applicationConfigurationDescription_zeppelinApplicationConfigurationDescription :: Lens' ApplicationConfigurationDescription (Maybe ZeppelinApplicationConfigurationDescription)

The configuration parameters for a Kinesis Data Analytics Studio notebook.

applicationConfigurationDescription_vpcConfigurationDescriptions :: Lens' ApplicationConfigurationDescription (Maybe [VpcConfigurationDescription])

The array of descriptions of VPC configurations available to the application.

applicationConfigurationDescription_applicationCodeConfigurationDescription :: Lens' ApplicationConfigurationDescription (Maybe ApplicationCodeConfigurationDescription)

The details about the application code for a Flink-based Kinesis Data Analytics application.

applicationConfigurationDescription_environmentPropertyDescriptions :: Lens' ApplicationConfigurationDescription (Maybe EnvironmentPropertyDescriptions)

Describes execution properties for a Flink-based Kinesis Data Analytics application.

applicationConfigurationDescription_runConfigurationDescription :: Lens' ApplicationConfigurationDescription (Maybe RunConfigurationDescription)

The details about the starting properties for a Kinesis Data Analytics application.

applicationConfigurationDescription_sqlApplicationConfigurationDescription :: Lens' ApplicationConfigurationDescription (Maybe SqlApplicationConfigurationDescription)

The details about inputs, outputs, and reference data sources for a SQL-based Kinesis Data Analytics application.

applicationConfigurationDescription_flinkApplicationConfigurationDescription :: Lens' ApplicationConfigurationDescription (Maybe FlinkApplicationConfigurationDescription)

The details about a Flink-based Kinesis Data Analytics application.

applicationConfigurationDescription_applicationSnapshotConfigurationDescription :: Lens' ApplicationConfigurationDescription (Maybe ApplicationSnapshotConfigurationDescription)

Describes whether snapshots are enabled for a Flink-based Kinesis Data Analytics application.