Network.AWS.KinesisAnalyticsV2.Types.ApplicationDetail

Description

Documentation

data ApplicationDetail

Describes the application, including the application Amazon Resource Name (ARN), status, latest version, and input and output configurations.

See: newApplicationDetail smart constructor.

Constructors

ApplicationDetail'

Fields

applicationDescription :: Maybe Text

The description of the application.

conditionalToken :: Maybe Text

A value you use to implement strong concurrency for application updates.

cloudWatchLoggingOptionDescriptions :: Maybe [CloudWatchLoggingOptionDescription]

Describes the application Amazon CloudWatch logging options.

applicationVersionRolledBackFrom :: Maybe Natural

If you reverted the application using RollbackApplication, the application version when RollbackApplication was called.

applicationMode :: Maybe ApplicationMode

To create a Kinesis Data Analytics Studio notebook, you must set the mode to INTERACTIVE. However, for a Kinesis Data Analytics for Apache Flink application, the mode is optional.

applicationConfigurationDescription :: Maybe ApplicationConfigurationDescription

Describes details about the application code and starting parameters for a Kinesis Data Analytics application.

applicationVersionRolledBackTo :: Maybe Natural

The version to which you want to roll back the application.

createTimestamp :: Maybe POSIX

The current timestamp when the application was created.

lastUpdateTimestamp :: Maybe POSIX

The current timestamp when the application was last updated.

serviceExecutionRole :: Maybe Text

Specifies the IAM role that the application uses to access external resources.

applicationVersionUpdatedFrom :: Maybe Natural

The previous application version before the latest application update. RollbackApplication reverts the application to this version.

applicationMaintenanceConfigurationDescription :: Maybe ApplicationMaintenanceConfigurationDescription

The details of the maintenance configuration for the application.

applicationARN :: Text

The ARN of the application.

applicationName :: Text

The name of the application.

runtimeEnvironment :: RuntimeEnvironment

The runtime environment for the application (SQL-1_0, FLINK-1_6, FLINK-1_8, or FLINK-1_11).

applicationStatus :: ApplicationStatus

The status of the application.

applicationVersionId :: Natural

Provides the current application version. Kinesis Data Analytics updates the ApplicationVersionId each time you update the application.

newApplicationDetail

Arguments

:: Text
$sel:applicationARN:ApplicationDetail'

-> Text
$sel:applicationName:ApplicationDetail'

-> RuntimeEnvironment
$sel:runtimeEnvironment:ApplicationDetail'

-> ApplicationStatus
$sel:applicationStatus:ApplicationDetail'

-> Natural
$sel:applicationVersionId:ApplicationDetail'

-> ApplicationDetail

Create a value of ApplicationDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationDescription:ApplicationDetail', applicationDetail_applicationDescription - The description of the application.

$sel:conditionalToken:ApplicationDetail', applicationDetail_conditionalToken - A value you use to implement strong concurrency for application updates.

$sel:cloudWatchLoggingOptionDescriptions:ApplicationDetail', applicationDetail_cloudWatchLoggingOptionDescriptions - Describes the application Amazon CloudWatch logging options.

$sel:applicationVersionRolledBackFrom:ApplicationDetail', applicationDetail_applicationVersionRolledBackFrom - If you reverted the application using RollbackApplication, the application version when RollbackApplication was called.

$sel:applicationMode:ApplicationDetail', applicationDetail_applicationMode - To create a Kinesis Data Analytics Studio notebook, you must set the mode to INTERACTIVE. However, for a Kinesis Data Analytics for Apache Flink application, the mode is optional.

$sel:applicationConfigurationDescription:ApplicationDetail', applicationDetail_applicationConfigurationDescription - Describes details about the application code and starting parameters for a Kinesis Data Analytics application.

$sel:applicationVersionRolledBackTo:ApplicationDetail', applicationDetail_applicationVersionRolledBackTo - The version to which you want to roll back the application.

$sel:createTimestamp:ApplicationDetail', applicationDetail_createTimestamp - The current timestamp when the application was created.

$sel:lastUpdateTimestamp:ApplicationDetail', applicationDetail_lastUpdateTimestamp - The current timestamp when the application was last updated.

$sel:serviceExecutionRole:ApplicationDetail', applicationDetail_serviceExecutionRole - Specifies the IAM role that the application uses to access external resources.

$sel:applicationVersionUpdatedFrom:ApplicationDetail', applicationDetail_applicationVersionUpdatedFrom - The previous application version before the latest application update. RollbackApplication reverts the application to this version.

$sel:applicationMaintenanceConfigurationDescription:ApplicationDetail', applicationDetail_applicationMaintenanceConfigurationDescription - The details of the maintenance configuration for the application.

$sel:applicationARN:ApplicationDetail', applicationDetail_applicationARN - The ARN of the application.

$sel:applicationName:ApplicationDetail', applicationDetail_applicationName - The name of the application.

$sel:runtimeEnvironment:ApplicationDetail', applicationDetail_runtimeEnvironment - The runtime environment for the application (SQL-1_0, FLINK-1_6, FLINK-1_8, or FLINK-1_11).

$sel:applicationStatus:ApplicationDetail', applicationDetail_applicationStatus - The status of the application.

$sel:applicationVersionId:ApplicationDetail', applicationDetail_applicationVersionId - Provides the current application version. Kinesis Data Analytics updates the ApplicationVersionId each time you update the application.

applicationDetail_applicationDescription :: Lens' ApplicationDetail (Maybe Text)

The description of the application.

applicationDetail_conditionalToken :: Lens' ApplicationDetail (Maybe Text)

A value you use to implement strong concurrency for application updates.

applicationDetail_cloudWatchLoggingOptionDescriptions :: Lens' ApplicationDetail (Maybe [CloudWatchLoggingOptionDescription])

Describes the application Amazon CloudWatch logging options.

applicationDetail_applicationVersionRolledBackFrom :: Lens' ApplicationDetail (Maybe Natural)

If you reverted the application using RollbackApplication, the application version when RollbackApplication was called.

applicationDetail_applicationMode :: Lens' ApplicationDetail (Maybe ApplicationMode)

To create a Kinesis Data Analytics Studio notebook, you must set the mode to INTERACTIVE. However, for a Kinesis Data Analytics for Apache Flink application, the mode is optional.

applicationDetail_applicationConfigurationDescription :: Lens' ApplicationDetail (Maybe ApplicationConfigurationDescription)

Describes details about the application code and starting parameters for a Kinesis Data Analytics application.

applicationDetail_applicationVersionRolledBackTo :: Lens' ApplicationDetail (Maybe Natural)

The version to which you want to roll back the application.

applicationDetail_createTimestamp :: Lens' ApplicationDetail (Maybe UTCTime)

The current timestamp when the application was created.

applicationDetail_lastUpdateTimestamp :: Lens' ApplicationDetail (Maybe UTCTime)

The current timestamp when the application was last updated.

applicationDetail_serviceExecutionRole :: Lens' ApplicationDetail (Maybe Text)

Specifies the IAM role that the application uses to access external resources.

applicationDetail_applicationVersionUpdatedFrom :: Lens' ApplicationDetail (Maybe Natural)

The previous application version before the latest application update. RollbackApplication reverts the application to this version.

applicationDetail_applicationMaintenanceConfigurationDescription :: Lens' ApplicationDetail (Maybe ApplicationMaintenanceConfigurationDescription)

The details of the maintenance configuration for the application.

applicationDetail_applicationARN :: Lens' ApplicationDetail Text

The ARN of the application.

applicationDetail_applicationName :: Lens' ApplicationDetail Text

The name of the application.

applicationDetail_runtimeEnvironment :: Lens' ApplicationDetail RuntimeEnvironment

The runtime environment for the application (SQL-1_0, FLINK-1_6, FLINK-1_8, or FLINK-1_11).

applicationDetail_applicationStatus :: Lens' ApplicationDetail ApplicationStatus

The status of the application.

applicationDetail_applicationVersionId :: Lens' ApplicationDetail Natural

Provides the current application version. Kinesis Data Analytics updates the ApplicationVersionId each time you update the application.