Network.AWS.KinesisAnalyticsV2.Types.CatalogConfigurationDescription

Description

Documentation

data CatalogConfigurationDescription

The configuration parameters for the default Amazon Glue database. You use this database for Apache Flink SQL queries and table API transforms that you write in a Kinesis Data Analytics Studio notebook.

See: newCatalogConfigurationDescription smart constructor.

Constructors

CatalogConfigurationDescription'

Fields

glueDataCatalogConfigurationDescription :: GlueDataCatalogConfigurationDescription

The configuration parameters for the default Amazon Glue database. You use this database for SQL queries that you write in a Kinesis Data Analytics Studio notebook.

newCatalogConfigurationDescription

Arguments

:: GlueDataCatalogConfigurationDescription
$sel:glueDataCatalogConfigurationDescription:CatalogConfigurationDescription'

-> CatalogConfigurationDescription

Create a value of CatalogConfigurationDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:glueDataCatalogConfigurationDescription:CatalogConfigurationDescription', catalogConfigurationDescription_glueDataCatalogConfigurationDescription - The configuration parameters for the default Amazon Glue database. You use this database for SQL queries that you write in a Kinesis Data Analytics Studio notebook.

catalogConfigurationDescription_glueDataCatalogConfigurationDescription :: Lens' CatalogConfigurationDescription GlueDataCatalogConfigurationDescription

The configuration parameters for the default Amazon Glue database. You use this database for SQL queries that you write in a Kinesis Data Analytics Studio notebook.