Network.AWS.KinesisAnalyticsV2.Types.CheckpointConfigurationDescription

Description

Documentation

data CheckpointConfigurationDescription

Describes checkpointing parameters for a Flink-based Kinesis Data Analytics application.

See: newCheckpointConfigurationDescription smart constructor.

Constructors

CheckpointConfigurationDescription'

Fields

checkpointingEnabled :: Maybe Bool

Describes whether checkpointing is enabled for a Flink-based Kinesis Data Analytics application.

If CheckpointConfiguration.ConfigurationType is DEFAULT, the application will use a CheckpointingEnabled value of true, even if this value is set to another value using this API or in application code.

configurationType :: Maybe ConfigurationType

Describes whether the application uses the default checkpointing behavior in Kinesis Data Analytics.

If this value is set to DEFAULT, the application will use the following values, even if they are set to other values using APIs or application code:

CheckpointingEnabled: true

CheckpointInterval: 60000

MinPauseBetweenCheckpoints: 5000

minPauseBetweenCheckpoints :: Maybe Natural

Describes the minimum time in milliseconds after a checkpoint operation completes that a new checkpoint operation can start.

If CheckpointConfiguration.ConfigurationType is DEFAULT, the application will use a MinPauseBetweenCheckpoints value of 5000, even if this value is set using this API or in application code.

checkpointInterval :: Maybe Natural

Describes the interval in milliseconds between checkpoint operations.

If CheckpointConfiguration.ConfigurationType is DEFAULT, the application will use a CheckpointInterval value of 60000, even if this value is set to another value using this API or in application code.

newCheckpointConfigurationDescription :: CheckpointConfigurationDescription

Create a value of CheckpointConfigurationDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checkpointingEnabled:CheckpointConfigurationDescription', checkpointConfigurationDescription_checkpointingEnabled - Describes whether checkpointing is enabled for a Flink-based Kinesis Data Analytics application.

If CheckpointConfiguration.ConfigurationType is DEFAULT, the application will use a CheckpointingEnabled value of true, even if this value is set to another value using this API or in application code.

$sel:configurationType:CheckpointConfigurationDescription', checkpointConfigurationDescription_configurationType - Describes whether the application uses the default checkpointing behavior in Kinesis Data Analytics.

If this value is set to DEFAULT, the application will use the following values, even if they are set to other values using APIs or application code:

CheckpointingEnabled: true

CheckpointInterval: 60000

MinPauseBetweenCheckpoints: 5000

$sel:minPauseBetweenCheckpoints:CheckpointConfigurationDescription', checkpointConfigurationDescription_minPauseBetweenCheckpoints - Describes the minimum time in milliseconds after a checkpoint operation completes that a new checkpoint operation can start.

If CheckpointConfiguration.ConfigurationType is DEFAULT, the application will use a MinPauseBetweenCheckpoints value of 5000, even if this value is set using this API or in application code.

$sel:checkpointInterval:CheckpointConfigurationDescription', checkpointConfigurationDescription_checkpointInterval - Describes the interval in milliseconds between checkpoint operations.

If CheckpointConfiguration.ConfigurationType is DEFAULT, the application will use a CheckpointInterval value of 60000, even if this value is set to another value using this API or in application code.

checkpointConfigurationDescription_checkpointingEnabled :: Lens' CheckpointConfigurationDescription (Maybe Bool)

Describes whether checkpointing is enabled for a Flink-based Kinesis Data Analytics application.

If CheckpointConfiguration.ConfigurationType is DEFAULT, the application will use a CheckpointingEnabled value of true, even if this value is set to another value using this API or in application code.

checkpointConfigurationDescription_configurationType :: Lens' CheckpointConfigurationDescription (Maybe ConfigurationType)

Describes whether the application uses the default checkpointing behavior in Kinesis Data Analytics.

If this value is set to DEFAULT, the application will use the following values, even if they are set to other values using APIs or application code:

CheckpointingEnabled: true

CheckpointInterval: 60000

MinPauseBetweenCheckpoints: 5000

checkpointConfigurationDescription_minPauseBetweenCheckpoints :: Lens' CheckpointConfigurationDescription (Maybe Natural)

Describes the minimum time in milliseconds after a checkpoint operation completes that a new checkpoint operation can start.

If CheckpointConfiguration.ConfigurationType is DEFAULT, the application will use a MinPauseBetweenCheckpoints value of 5000, even if this value is set using this API or in application code.

checkpointConfigurationDescription_checkpointInterval :: Lens' CheckpointConfigurationDescription (Maybe Natural)

Describes the interval in milliseconds between checkpoint operations.

If CheckpointConfiguration.ConfigurationType is DEFAULT, the application will use a CheckpointInterval value of 60000, even if this value is set to another value using this API or in application code.