Network.AWS.KinesisAnalyticsV2.Types.CheckpointConfigurationUpdate

Description

Documentation

data CheckpointConfigurationUpdate

Describes updates to the checkpointing parameters for a Flink-based Kinesis Data Analytics application.

See: newCheckpointConfigurationUpdate smart constructor.

Constructors

CheckpointConfigurationUpdate'

Fields

checkpointIntervalUpdate :: Maybe Natural

Describes updates to the interval in milliseconds between checkpoint operations.

If CheckpointConfiguration.ConfigurationType is DEFAULT, the application will use a CheckpointInterval value of 60000, even if this value is set to another value using this API or in application code.

checkpointingEnabledUpdate :: Maybe Bool

Describes updates to whether checkpointing is enabled for an application.

If CheckpointConfiguration.ConfigurationType is DEFAULT, the application will use a CheckpointingEnabled value of true, even if this value is set to another value using this API or in application code.

minPauseBetweenCheckpointsUpdate :: Maybe Natural

Describes updates to the minimum time in milliseconds after a checkpoint operation completes that a new checkpoint operation can start.

If CheckpointConfiguration.ConfigurationType is DEFAULT, the application will use a MinPauseBetweenCheckpoints value of 5000, even if this value is set using this API or in application code.

configurationTypeUpdate :: Maybe ConfigurationType

Describes updates to whether the application uses the default checkpointing behavior of Kinesis Data Analytics. You must set this property to CUSTOM in order to set the CheckpointingEnabled, CheckpointInterval, or MinPauseBetweenCheckpoints parameters.

If this value is set to DEFAULT, the application will use the following values, even if they are set to other values using APIs or application code:

CheckpointingEnabled: true

CheckpointInterval: 60000

MinPauseBetweenCheckpoints: 5000

newCheckpointConfigurationUpdate :: CheckpointConfigurationUpdate

Create a value of CheckpointConfigurationUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checkpointIntervalUpdate:CheckpointConfigurationUpdate', checkpointConfigurationUpdate_checkpointIntervalUpdate - Describes updates to the interval in milliseconds between checkpoint operations.

If CheckpointConfiguration.ConfigurationType is DEFAULT, the application will use a CheckpointInterval value of 60000, even if this value is set to another value using this API or in application code.

$sel:checkpointingEnabledUpdate:CheckpointConfigurationUpdate', checkpointConfigurationUpdate_checkpointingEnabledUpdate - Describes updates to whether checkpointing is enabled for an application.

If CheckpointConfiguration.ConfigurationType is DEFAULT, the application will use a CheckpointingEnabled value of true, even if this value is set to another value using this API or in application code.

$sel:minPauseBetweenCheckpointsUpdate:CheckpointConfigurationUpdate', checkpointConfigurationUpdate_minPauseBetweenCheckpointsUpdate - Describes updates to the minimum time in milliseconds after a checkpoint operation completes that a new checkpoint operation can start.

If CheckpointConfiguration.ConfigurationType is DEFAULT, the application will use a MinPauseBetweenCheckpoints value of 5000, even if this value is set using this API or in application code.

$sel:configurationTypeUpdate:CheckpointConfigurationUpdate', checkpointConfigurationUpdate_configurationTypeUpdate - Describes updates to whether the application uses the default checkpointing behavior of Kinesis Data Analytics. You must set this property to CUSTOM in order to set the CheckpointingEnabled, CheckpointInterval, or MinPauseBetweenCheckpoints parameters.

If this value is set to DEFAULT, the application will use the following values, even if they are set to other values using APIs or application code:

CheckpointingEnabled: true

CheckpointInterval: 60000

MinPauseBetweenCheckpoints: 5000

checkpointConfigurationUpdate_checkpointIntervalUpdate :: Lens' CheckpointConfigurationUpdate (Maybe Natural)

Describes updates to the interval in milliseconds between checkpoint operations.

If CheckpointConfiguration.ConfigurationType is DEFAULT, the application will use a CheckpointInterval value of 60000, even if this value is set to another value using this API or in application code.

checkpointConfigurationUpdate_checkpointingEnabledUpdate :: Lens' CheckpointConfigurationUpdate (Maybe Bool)

Describes updates to whether checkpointing is enabled for an application.

If CheckpointConfiguration.ConfigurationType is DEFAULT, the application will use a CheckpointingEnabled value of true, even if this value is set to another value using this API or in application code.

checkpointConfigurationUpdate_minPauseBetweenCheckpointsUpdate :: Lens' CheckpointConfigurationUpdate (Maybe Natural)

Describes updates to the minimum time in milliseconds after a checkpoint operation completes that a new checkpoint operation can start.

If CheckpointConfiguration.ConfigurationType is DEFAULT, the application will use a MinPauseBetweenCheckpoints value of 5000, even if this value is set using this API or in application code.

checkpointConfigurationUpdate_configurationTypeUpdate :: Lens' CheckpointConfigurationUpdate (Maybe ConfigurationType)

Describes updates to whether the application uses the default checkpointing behavior of Kinesis Data Analytics. You must set this property to CUSTOM in order to set the CheckpointingEnabled, CheckpointInterval, or MinPauseBetweenCheckpoints parameters.

If this value is set to DEFAULT, the application will use the following values, even if they are set to other values using APIs or application code:

CheckpointingEnabled: true

CheckpointInterval: 60000

MinPauseBetweenCheckpoints: 5000