Network.AWS.KinesisAnalyticsV2.Types.CloudWatchLoggingOptionDescription

Description

Documentation

data CloudWatchLoggingOptionDescription

Describes the Amazon CloudWatch logging option.

See: newCloudWatchLoggingOptionDescription smart constructor.

Constructors

CloudWatchLoggingOptionDescription'

Fields

cloudWatchLoggingOptionId :: Maybe Text

The ID of the CloudWatch logging option description.

roleARN :: Maybe Text

The IAM ARN of the role to use to send application messages.

Provided for backward compatibility. Applications created with the current API version have an application-level service execution role rather than a resource-level role.

logStreamARN :: Text

The Amazon Resource Name (ARN) of the CloudWatch log to receive application messages.

newCloudWatchLoggingOptionDescription

Arguments

:: Text
$sel:logStreamARN:CloudWatchLoggingOptionDescription'

-> CloudWatchLoggingOptionDescription

Create a value of CloudWatchLoggingOptionDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cloudWatchLoggingOptionId:CloudWatchLoggingOptionDescription', cloudWatchLoggingOptionDescription_cloudWatchLoggingOptionId - The ID of the CloudWatch logging option description.

$sel:roleARN:CloudWatchLoggingOptionDescription', cloudWatchLoggingOptionDescription_roleARN - The IAM ARN of the role to use to send application messages.

Provided for backward compatibility. Applications created with the current API version have an application-level service execution role rather than a resource-level role.

$sel:logStreamARN:CloudWatchLoggingOptionDescription', cloudWatchLoggingOptionDescription_logStreamARN - The Amazon Resource Name (ARN) of the CloudWatch log to receive application messages.

cloudWatchLoggingOptionDescription_cloudWatchLoggingOptionId :: Lens' CloudWatchLoggingOptionDescription (Maybe Text)

The ID of the CloudWatch logging option description.

cloudWatchLoggingOptionDescription_roleARN :: Lens' CloudWatchLoggingOptionDescription (Maybe Text)

The IAM ARN of the role to use to send application messages.

Provided for backward compatibility. Applications created with the current API version have an application-level service execution role rather than a resource-level role.

cloudWatchLoggingOptionDescription_logStreamARN :: Lens' CloudWatchLoggingOptionDescription Text

The Amazon Resource Name (ARN) of the CloudWatch log to receive application messages.