Network.AWS.KinesisAnalyticsV2.Types.FlinkApplicationConfiguration

Description

Documentation

data FlinkApplicationConfiguration

Describes configuration parameters for a Flink-based Kinesis Data Analytics application or a Studio notebook.

See: newFlinkApplicationConfiguration smart constructor.

Constructors

FlinkApplicationConfiguration'

Fields

monitoringConfiguration :: Maybe MonitoringConfiguration

Describes configuration parameters for Amazon CloudWatch logging for an application.

parallelismConfiguration :: Maybe ParallelismConfiguration

Describes parameters for how an application executes multiple tasks simultaneously.

checkpointConfiguration :: Maybe CheckpointConfiguration

Describes an application's checkpointing configuration. Checkpointing is the process of persisting application state for fault tolerance. For more information, see Checkpoints for Fault Tolerance in the Apache Flink Documentation.

newFlinkApplicationConfiguration :: FlinkApplicationConfiguration

Create a value of FlinkApplicationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:monitoringConfiguration:FlinkApplicationConfiguration', flinkApplicationConfiguration_monitoringConfiguration - Describes configuration parameters for Amazon CloudWatch logging for an application.

$sel:parallelismConfiguration:FlinkApplicationConfiguration', flinkApplicationConfiguration_parallelismConfiguration - Describes parameters for how an application executes multiple tasks simultaneously.

$sel:checkpointConfiguration:FlinkApplicationConfiguration', flinkApplicationConfiguration_checkpointConfiguration - Describes an application's checkpointing configuration. Checkpointing is the process of persisting application state for fault tolerance. For more information, see Checkpoints for Fault Tolerance in the Apache Flink Documentation.

flinkApplicationConfiguration_monitoringConfiguration :: Lens' FlinkApplicationConfiguration (Maybe MonitoringConfiguration)

Describes configuration parameters for Amazon CloudWatch logging for an application.

flinkApplicationConfiguration_parallelismConfiguration :: Lens' FlinkApplicationConfiguration (Maybe ParallelismConfiguration)

Describes parameters for how an application executes multiple tasks simultaneously.

flinkApplicationConfiguration_checkpointConfiguration :: Lens' FlinkApplicationConfiguration (Maybe CheckpointConfiguration)

Describes an application's checkpointing configuration. Checkpointing is the process of persisting application state for fault tolerance. For more information, see Checkpoints for Fault Tolerance in the Apache Flink Documentation.