Network.AWS.KinesisAnalyticsV2.Types.GlueDataCatalogConfigurationUpdate

Description

Documentation

data GlueDataCatalogConfigurationUpdate

Updates to the configuration of the Glue Data Catalog that you use for SQL queries that you write in a Kinesis Data Analytics Studio notebook.

See: newGlueDataCatalogConfigurationUpdate smart constructor.

Constructors

GlueDataCatalogConfigurationUpdate'

Fields

databaseARNUpdate :: Text

The updated Amazon Resource Name (ARN) of the database.

newGlueDataCatalogConfigurationUpdate

Arguments

:: Text
$sel:databaseARNUpdate:GlueDataCatalogConfigurationUpdate'

-> GlueDataCatalogConfigurationUpdate

Create a value of GlueDataCatalogConfigurationUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:databaseARNUpdate:GlueDataCatalogConfigurationUpdate', glueDataCatalogConfigurationUpdate_databaseARNUpdate - The updated Amazon Resource Name (ARN) of the database.

glueDataCatalogConfigurationUpdate_databaseARNUpdate :: Lens' GlueDataCatalogConfigurationUpdate Text

The updated Amazon Resource Name (ARN) of the database.