Network.AWS.KinesisAnalyticsV2.Types.Input

Description

Documentation

data Input

When you configure the application input for a SQL-based Kinesis Data Analytics application, you specify the streaming source, the in-application stream name that is created, and the mapping between the two.

See: newInput smart constructor.

Constructors

Input'

Fields

inputParallelism :: Maybe InputParallelism

Describes the number of in-application streams to create.

inputProcessingConfiguration :: Maybe InputProcessingConfiguration

The InputProcessingConfiguration for the input. An input processor transforms records as they are received from the stream, before the application's SQL code executes. Currently, the only input processing configuration available is InputLambdaProcessor.

kinesisStreamsInput :: Maybe KinesisStreamsInput

If the streaming source is an Amazon Kinesis data stream, identifies the stream's Amazon Resource Name (ARN).

kinesisFirehoseInput :: Maybe KinesisFirehoseInput

If the streaming source is an Amazon Kinesis Data Firehose delivery stream, identifies the delivery stream's ARN.

namePrefix :: Text

The name prefix to use when creating an in-application stream. Suppose that you specify a prefix "MyInApplicationStream." Kinesis Data Analytics then creates one or more (as per the InputParallelism count you specified) in-application streams with the names "MyInApplicationStream_001," "MyInApplicationStream_002," and so on.

inputSchema :: SourceSchema

Describes the format of the data in the streaming source, and how each data element maps to corresponding columns in the in-application stream that is being created.

Also used to describe the format of the reference data source.

newInput

Arguments

:: Text
$sel:namePrefix:Input'

-> SourceSchema
$sel:inputSchema:Input'

-> Input

Create a value of Input with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputParallelism:Input', input_inputParallelism - Describes the number of in-application streams to create.

$sel:inputProcessingConfiguration:Input', input_inputProcessingConfiguration - The InputProcessingConfiguration for the input. An input processor transforms records as they are received from the stream, before the application's SQL code executes. Currently, the only input processing configuration available is InputLambdaProcessor.

$sel:kinesisStreamsInput:Input', input_kinesisStreamsInput - If the streaming source is an Amazon Kinesis data stream, identifies the stream's Amazon Resource Name (ARN).

$sel:kinesisFirehoseInput:Input', input_kinesisFirehoseInput - If the streaming source is an Amazon Kinesis Data Firehose delivery stream, identifies the delivery stream's ARN.

$sel:namePrefix:Input', input_namePrefix - The name prefix to use when creating an in-application stream. Suppose that you specify a prefix "MyInApplicationStream." Kinesis Data Analytics then creates one or more (as per the InputParallelism count you specified) in-application streams with the names "MyInApplicationStream_001," "MyInApplicationStream_002," and so on.

$sel:inputSchema:Input', input_inputSchema - Describes the format of the data in the streaming source, and how each data element maps to corresponding columns in the in-application stream that is being created.

Also used to describe the format of the reference data source.

input_inputParallelism :: Lens' Input (Maybe InputParallelism)

Describes the number of in-application streams to create.

input_inputProcessingConfiguration :: Lens' Input (Maybe InputProcessingConfiguration)

The InputProcessingConfiguration for the input. An input processor transforms records as they are received from the stream, before the application's SQL code executes. Currently, the only input processing configuration available is InputLambdaProcessor.

input_kinesisStreamsInput :: Lens' Input (Maybe KinesisStreamsInput)

If the streaming source is an Amazon Kinesis data stream, identifies the stream's Amazon Resource Name (ARN).

input_kinesisFirehoseInput :: Lens' Input (Maybe KinesisFirehoseInput)

If the streaming source is an Amazon Kinesis Data Firehose delivery stream, identifies the delivery stream's ARN.

input_namePrefix :: Lens' Input Text

The name prefix to use when creating an in-application stream. Suppose that you specify a prefix "MyInApplicationStream." Kinesis Data Analytics then creates one or more (as per the InputParallelism count you specified) in-application streams with the names "MyInApplicationStream_001," "MyInApplicationStream_002," and so on.

input_inputSchema :: Lens' Input SourceSchema

Describes the format of the data in the streaming source, and how each data element maps to corresponding columns in the in-application stream that is being created.

Also used to describe the format of the reference data source.