Network.AWS.KinesisAnalyticsV2.Types.InputDescription

Description

Documentation

data InputDescription

Describes the application input configuration for a SQL-based Kinesis Data Analytics application.

See: newInputDescription smart constructor.

Constructors

InputDescription'

Fields

inputStartingPositionConfiguration :: Maybe InputStartingPositionConfiguration

The point at which the application is configured to read from the input stream.

inputParallelism :: Maybe InputParallelism

Describes the configured parallelism (number of in-application streams mapped to the streaming source).

inputId :: Maybe Text

The input ID that is associated with the application input. This is the ID that Kinesis Data Analytics assigns to each input configuration that you add to your application.

inAppStreamNames :: Maybe [Text]

Returns the in-application stream names that are mapped to the stream source.

kinesisFirehoseInputDescription :: Maybe KinesisFirehoseInputDescription

If a Kinesis Data Firehose delivery stream is configured as a streaming source, provides the delivery stream's ARN.

inputSchema :: Maybe SourceSchema

Describes the format of the data in the streaming source, and how each data element maps to corresponding columns in the in-application stream that is being created.

kinesisStreamsInputDescription :: Maybe KinesisStreamsInputDescription

If a Kinesis data stream is configured as a streaming source, provides the Kinesis data stream's Amazon Resource Name (ARN).

namePrefix :: Maybe Text

The in-application name prefix.

inputProcessingConfigurationDescription :: Maybe InputProcessingConfigurationDescription

The description of the preprocessor that executes on records in this input before the application's code is run.

newInputDescription :: InputDescription

Create a value of InputDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputStartingPositionConfiguration:InputDescription', inputDescription_inputStartingPositionConfiguration - The point at which the application is configured to read from the input stream.

$sel:inputParallelism:InputDescription', inputDescription_inputParallelism - Describes the configured parallelism (number of in-application streams mapped to the streaming source).

$sel:inputId:InputDescription', inputDescription_inputId - The input ID that is associated with the application input. This is the ID that Kinesis Data Analytics assigns to each input configuration that you add to your application.

$sel:inAppStreamNames:InputDescription', inputDescription_inAppStreamNames - Returns the in-application stream names that are mapped to the stream source.

$sel:kinesisFirehoseInputDescription:InputDescription', inputDescription_kinesisFirehoseInputDescription - If a Kinesis Data Firehose delivery stream is configured as a streaming source, provides the delivery stream's ARN.

$sel:inputSchema:InputDescription', inputDescription_inputSchema - Describes the format of the data in the streaming source, and how each data element maps to corresponding columns in the in-application stream that is being created.

$sel:kinesisStreamsInputDescription:InputDescription', inputDescription_kinesisStreamsInputDescription - If a Kinesis data stream is configured as a streaming source, provides the Kinesis data stream's Amazon Resource Name (ARN).

$sel:namePrefix:InputDescription', inputDescription_namePrefix - The in-application name prefix.

$sel:inputProcessingConfigurationDescription:InputDescription', inputDescription_inputProcessingConfigurationDescription - The description of the preprocessor that executes on records in this input before the application's code is run.

inputDescription_inputStartingPositionConfiguration :: Lens' InputDescription (Maybe InputStartingPositionConfiguration)

The point at which the application is configured to read from the input stream.

inputDescription_inputParallelism :: Lens' InputDescription (Maybe InputParallelism)

Describes the configured parallelism (number of in-application streams mapped to the streaming source).

inputDescription_inputId :: Lens' InputDescription (Maybe Text)

The input ID that is associated with the application input. This is the ID that Kinesis Data Analytics assigns to each input configuration that you add to your application.

inputDescription_inAppStreamNames :: Lens' InputDescription (Maybe [Text])

Returns the in-application stream names that are mapped to the stream source.

inputDescription_kinesisFirehoseInputDescription :: Lens' InputDescription (Maybe KinesisFirehoseInputDescription)

If a Kinesis Data Firehose delivery stream is configured as a streaming source, provides the delivery stream's ARN.

inputDescription_inputSchema :: Lens' InputDescription (Maybe SourceSchema)

Describes the format of the data in the streaming source, and how each data element maps to corresponding columns in the in-application stream that is being created.

inputDescription_kinesisStreamsInputDescription :: Lens' InputDescription (Maybe KinesisStreamsInputDescription)

If a Kinesis data stream is configured as a streaming source, provides the Kinesis data stream's Amazon Resource Name (ARN).

inputDescription_namePrefix :: Lens' InputDescription (Maybe Text)

The in-application name prefix.

inputDescription_inputProcessingConfigurationDescription :: Lens' InputDescription (Maybe InputProcessingConfigurationDescription)

The description of the preprocessor that executes on records in this input before the application's code is run.