Network.AWS.KinesisAnalyticsV2.Types.InputUpdate

Description

Documentation

data InputUpdate

For a SQL-based Kinesis Data Analytics application, describes updates to a specific input configuration (identified by the InputId of an application).

See: newInputUpdate smart constructor.

Constructors

InputUpdate'

Fields

inputProcessingConfigurationUpdate :: Maybe InputProcessingConfigurationUpdate

Describes updates to an InputProcessingConfiguration.

kinesisStreamsInputUpdate :: Maybe KinesisStreamsInputUpdate

If a Kinesis data stream is the streaming source to be updated, provides an updated stream Amazon Resource Name (ARN).

inputParallelismUpdate :: Maybe InputParallelismUpdate

Describes the parallelism updates (the number of in-application streams Kinesis Data Analytics creates for the specific streaming source).

namePrefixUpdate :: Maybe Text

The name prefix for in-application streams that Kinesis Data Analytics creates for the specific streaming source.

inputSchemaUpdate :: Maybe InputSchemaUpdate

Describes the data format on the streaming source, and how record elements on the streaming source map to columns of the in-application stream that is created.

kinesisFirehoseInputUpdate :: Maybe KinesisFirehoseInputUpdate

If a Kinesis Data Firehose delivery stream is the streaming source to be updated, provides an updated stream ARN.

inputId :: Text

The input ID of the application input to be updated.

newInputUpdate

Arguments

:: Text
$sel:inputId:InputUpdate'

-> InputUpdate

Create a value of InputUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputProcessingConfigurationUpdate:InputUpdate', inputUpdate_inputProcessingConfigurationUpdate - Describes updates to an InputProcessingConfiguration.

$sel:kinesisStreamsInputUpdate:InputUpdate', inputUpdate_kinesisStreamsInputUpdate - If a Kinesis data stream is the streaming source to be updated, provides an updated stream Amazon Resource Name (ARN).

$sel:inputParallelismUpdate:InputUpdate', inputUpdate_inputParallelismUpdate - Describes the parallelism updates (the number of in-application streams Kinesis Data Analytics creates for the specific streaming source).

$sel:namePrefixUpdate:InputUpdate', inputUpdate_namePrefixUpdate - The name prefix for in-application streams that Kinesis Data Analytics creates for the specific streaming source.

$sel:inputSchemaUpdate:InputUpdate', inputUpdate_inputSchemaUpdate - Describes the data format on the streaming source, and how record elements on the streaming source map to columns of the in-application stream that is created.

$sel:kinesisFirehoseInputUpdate:InputUpdate', inputUpdate_kinesisFirehoseInputUpdate - If a Kinesis Data Firehose delivery stream is the streaming source to be updated, provides an updated stream ARN.

$sel:inputId:InputUpdate', inputUpdate_inputId - The input ID of the application input to be updated.

inputUpdate_inputProcessingConfigurationUpdate :: Lens' InputUpdate (Maybe InputProcessingConfigurationUpdate)

Describes updates to an InputProcessingConfiguration.

inputUpdate_kinesisStreamsInputUpdate :: Lens' InputUpdate (Maybe KinesisStreamsInputUpdate)

If a Kinesis data stream is the streaming source to be updated, provides an updated stream Amazon Resource Name (ARN).

inputUpdate_inputParallelismUpdate :: Lens' InputUpdate (Maybe InputParallelismUpdate)

Describes the parallelism updates (the number of in-application streams Kinesis Data Analytics creates for the specific streaming source).

inputUpdate_namePrefixUpdate :: Lens' InputUpdate (Maybe Text)

The name prefix for in-application streams that Kinesis Data Analytics creates for the specific streaming source.

inputUpdate_inputSchemaUpdate :: Lens' InputUpdate (Maybe InputSchemaUpdate)

Describes the data format on the streaming source, and how record elements on the streaming source map to columns of the in-application stream that is created.

inputUpdate_kinesisFirehoseInputUpdate :: Lens' InputUpdate (Maybe KinesisFirehoseInputUpdate)

If a Kinesis Data Firehose delivery stream is the streaming source to be updated, provides an updated stream ARN.

inputUpdate_inputId :: Lens' InputUpdate Text

The input ID of the application input to be updated.