Network.AWS.KinesisAnalyticsV2.Types.KinesisStreamsOutputDescription

Description

Documentation

data KinesisStreamsOutputDescription

For an SQL-based Kinesis Data Analytics application's output, describes the Kinesis data stream that is configured as its destination.

See: newKinesisStreamsOutputDescription smart constructor.

Constructors

KinesisStreamsOutputDescription'

Fields

roleARN :: Maybe Text

The ARN of the IAM role that Kinesis Data Analytics can assume to access the stream.

Provided for backward compatibility. Applications that are created with the current API version have an application-level service execution role rather than a resource-level role.

resourceARN :: Text

The Amazon Resource Name (ARN) of the Kinesis data stream.

newKinesisStreamsOutputDescription

Arguments

:: Text
$sel:resourceARN:KinesisStreamsOutputDescription'

-> KinesisStreamsOutputDescription

Create a value of KinesisStreamsOutputDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleARN:KinesisStreamsOutputDescription', kinesisStreamsOutputDescription_roleARN - The ARN of the IAM role that Kinesis Data Analytics can assume to access the stream.

Provided for backward compatibility. Applications that are created with the current API version have an application-level service execution role rather than a resource-level role.

$sel:resourceARN:KinesisStreamsOutputDescription', kinesisStreamsOutputDescription_resourceARN - The Amazon Resource Name (ARN) of the Kinesis data stream.

kinesisStreamsOutputDescription_roleARN :: Lens' KinesisStreamsOutputDescription (Maybe Text)

The ARN of the IAM role that Kinesis Data Analytics can assume to access the stream.

Provided for backward compatibility. Applications that are created with the current API version have an application-level service execution role rather than a resource-level role.

kinesisStreamsOutputDescription_resourceARN :: Lens' KinesisStreamsOutputDescription Text

The Amazon Resource Name (ARN) of the Kinesis data stream.