Network.AWS.KinesisAnalyticsV2.Types.MonitoringConfigurationUpdate

Description

Documentation

data MonitoringConfigurationUpdate

Describes updates to configuration parameters for Amazon CloudWatch logging for an application.

See: newMonitoringConfigurationUpdate smart constructor.

Constructors

MonitoringConfigurationUpdate'

Fields

metricsLevelUpdate :: Maybe MetricsLevel

Describes updates to the granularity of the CloudWatch Logs for an application. The Parallelism level is not recommended for applications with a Parallelism over 64 due to excessive costs.

logLevelUpdate :: Maybe LogLevel

Describes updates to the verbosity of the CloudWatch Logs for an application.

configurationTypeUpdate :: Maybe ConfigurationType

Describes updates to whether to use the default CloudWatch logging configuration for an application. You must set this property to CUSTOM in order to set the LogLevel or MetricsLevel parameters.

newMonitoringConfigurationUpdate :: MonitoringConfigurationUpdate

Create a value of MonitoringConfigurationUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricsLevelUpdate:MonitoringConfigurationUpdate', monitoringConfigurationUpdate_metricsLevelUpdate - Describes updates to the granularity of the CloudWatch Logs for an application. The Parallelism level is not recommended for applications with a Parallelism over 64 due to excessive costs.

$sel:logLevelUpdate:MonitoringConfigurationUpdate', monitoringConfigurationUpdate_logLevelUpdate - Describes updates to the verbosity of the CloudWatch Logs for an application.

$sel:configurationTypeUpdate:MonitoringConfigurationUpdate', monitoringConfigurationUpdate_configurationTypeUpdate - Describes updates to whether to use the default CloudWatch logging configuration for an application. You must set this property to CUSTOM in order to set the LogLevel or MetricsLevel parameters.

monitoringConfigurationUpdate_metricsLevelUpdate :: Lens' MonitoringConfigurationUpdate (Maybe MetricsLevel)

Describes updates to the granularity of the CloudWatch Logs for an application. The Parallelism level is not recommended for applications with a Parallelism over 64 due to excessive costs.

monitoringConfigurationUpdate_logLevelUpdate :: Lens' MonitoringConfigurationUpdate (Maybe LogLevel)

Describes updates to the verbosity of the CloudWatch Logs for an application.

monitoringConfigurationUpdate_configurationTypeUpdate :: Lens' MonitoringConfigurationUpdate (Maybe ConfigurationType)

Describes updates to whether to use the default CloudWatch logging configuration for an application. You must set this property to CUSTOM in order to set the LogLevel or MetricsLevel parameters.