Network.AWS.KinesisAnalyticsV2.Types.Output

Description

Documentation

data Output

Describes a SQL-based Kinesis Data Analytics application's output configuration, in which you identify an in-application stream and a destination where you want the in-application stream data to be written. The destination can be a Kinesis data stream or a Kinesis Data Firehose delivery stream.

See: newOutput smart constructor.

Constructors

Output'

Fields

lambdaOutput :: Maybe LambdaOutput

Identifies an Amazon Lambda function as the destination.

kinesisStreamsOutput :: Maybe KinesisStreamsOutput

Identifies a Kinesis data stream as the destination.

kinesisFirehoseOutput :: Maybe KinesisFirehoseOutput

Identifies a Kinesis Data Firehose delivery stream as the destination.

name :: Text

The name of the in-application stream.

destinationSchema :: DestinationSchema

Describes the data format when records are written to the destination.

newOutput

Arguments

:: Text
$sel:name:Output'

-> DestinationSchema
$sel:destinationSchema:Output'

-> Output

Create a value of Output with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lambdaOutput:Output', output_lambdaOutput - Identifies an Amazon Lambda function as the destination.

$sel:kinesisStreamsOutput:Output', output_kinesisStreamsOutput - Identifies a Kinesis data stream as the destination.

$sel:kinesisFirehoseOutput:Output', output_kinesisFirehoseOutput - Identifies a Kinesis Data Firehose delivery stream as the destination.

$sel:name:Output', output_name - The name of the in-application stream.

$sel:destinationSchema:Output', output_destinationSchema - Describes the data format when records are written to the destination.

output_lambdaOutput :: Lens' Output (Maybe LambdaOutput)

Identifies an Amazon Lambda function as the destination.

output_kinesisStreamsOutput :: Lens' Output (Maybe KinesisStreamsOutput)

Identifies a Kinesis data stream as the destination.

output_kinesisFirehoseOutput :: Lens' Output (Maybe KinesisFirehoseOutput)

Identifies a Kinesis Data Firehose delivery stream as the destination.

output_name :: Lens' Output Text

The name of the in-application stream.

output_destinationSchema :: Lens' Output DestinationSchema

Describes the data format when records are written to the destination.