Network.AWS.KinesisAnalyticsV2.Types.SourceSchema

Description

Documentation

data SourceSchema

For a SQL-based Kinesis Data Analytics application, describes the format of the data in the streaming source, and how each data element maps to corresponding columns created in the in-application stream.

See: newSourceSchema smart constructor.

Constructors

SourceSchema'

Fields

recordEncoding :: Maybe Text

Specifies the encoding of the records in the streaming source. For example, UTF-8.

recordFormat :: RecordFormat

Specifies the format of the records on the streaming source.

recordColumns :: NonEmpty RecordColumn

A list of RecordColumn objects.

newSourceSchema

Arguments

:: RecordFormat
$sel:recordFormat:SourceSchema'

-> NonEmpty RecordColumn
$sel:recordColumns:SourceSchema'

-> SourceSchema

Create a value of SourceSchema with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recordEncoding:SourceSchema', sourceSchema_recordEncoding - Specifies the encoding of the records in the streaming source. For example, UTF-8.

$sel:recordFormat:SourceSchema', sourceSchema_recordFormat - Specifies the format of the records on the streaming source.

$sel:recordColumns:SourceSchema', sourceSchema_recordColumns - A list of RecordColumn objects.

sourceSchema_recordEncoding :: Lens' SourceSchema (Maybe Text)

Specifies the encoding of the records in the streaming source. For example, UTF-8.

sourceSchema_recordFormat :: Lens' SourceSchema RecordFormat

Specifies the format of the records on the streaming source.

sourceSchema_recordColumns :: Lens' SourceSchema (NonEmpty RecordColumn)

A list of RecordColumn objects.