Network.AWS.KinesisAnalyticsV2.Types.ZeppelinApplicationConfigurationDescription

Description

Documentation

data ZeppelinApplicationConfigurationDescription

The configuration of a Kinesis Data Analytics Studio notebook.

See: newZeppelinApplicationConfigurationDescription smart constructor.

Constructors

ZeppelinApplicationConfigurationDescription'

Fields

customArtifactsConfigurationDescription :: Maybe [CustomArtifactConfigurationDescription]

Custom artifacts are dependency JARs and user-defined functions (UDF).

deployAsApplicationConfigurationDescription :: Maybe DeployAsApplicationConfigurationDescription

The parameters required to deploy a Kinesis Data Analytics Studio notebook as an application with durable state.

catalogConfigurationDescription :: Maybe CatalogConfigurationDescription

The Amazon Glue Data Catalog that is associated with the Kinesis Data Analytics Studio notebook.

monitoringConfigurationDescription :: ZeppelinMonitoringConfigurationDescription

The monitoring configuration of a Kinesis Data Analytics Studio notebook.

newZeppelinApplicationConfigurationDescription

Arguments

:: ZeppelinMonitoringConfigurationDescription
$sel:monitoringConfigurationDescription:ZeppelinApplicationConfigurationDescription'

-> ZeppelinApplicationConfigurationDescription

Create a value of ZeppelinApplicationConfigurationDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customArtifactsConfigurationDescription:ZeppelinApplicationConfigurationDescription', zeppelinApplicationConfigurationDescription_customArtifactsConfigurationDescription - Custom artifacts are dependency JARs and user-defined functions (UDF).

$sel:deployAsApplicationConfigurationDescription:ZeppelinApplicationConfigurationDescription', zeppelinApplicationConfigurationDescription_deployAsApplicationConfigurationDescription - The parameters required to deploy a Kinesis Data Analytics Studio notebook as an application with durable state.

$sel:catalogConfigurationDescription:ZeppelinApplicationConfigurationDescription', zeppelinApplicationConfigurationDescription_catalogConfigurationDescription - The Amazon Glue Data Catalog that is associated with the Kinesis Data Analytics Studio notebook.

$sel:monitoringConfigurationDescription:ZeppelinApplicationConfigurationDescription', zeppelinApplicationConfigurationDescription_monitoringConfigurationDescription - The monitoring configuration of a Kinesis Data Analytics Studio notebook.

zeppelinApplicationConfigurationDescription_customArtifactsConfigurationDescription :: Lens' ZeppelinApplicationConfigurationDescription (Maybe [CustomArtifactConfigurationDescription])

Custom artifacts are dependency JARs and user-defined functions (UDF).

zeppelinApplicationConfigurationDescription_deployAsApplicationConfigurationDescription :: Lens' ZeppelinApplicationConfigurationDescription (Maybe DeployAsApplicationConfigurationDescription)

The parameters required to deploy a Kinesis Data Analytics Studio notebook as an application with durable state.

zeppelinApplicationConfigurationDescription_catalogConfigurationDescription :: Lens' ZeppelinApplicationConfigurationDescription (Maybe CatalogConfigurationDescription)

The Amazon Glue Data Catalog that is associated with the Kinesis Data Analytics Studio notebook.

zeppelinApplicationConfigurationDescription_monitoringConfigurationDescription :: Lens' ZeppelinApplicationConfigurationDescription ZeppelinMonitoringConfigurationDescription

The monitoring configuration of a Kinesis Data Analytics Studio notebook.