Network.AWS.KinesisAnalyticsV2.UpdateApplicationMaintenanceConfiguration

Description

Updates the maintenance configuration of the Kinesis Data Analytics application.

You can invoke this operation on an application that is in one of the two following states: READY or RUNNING. If you invoke it when the application is in a state other than these two states, it throws a ResourceInUseException. The service makes use of the updated configuration the next time it schedules maintenance for the application. If you invoke this operation after the service schedules maintenance, the service will apply the configuration update the next time it schedules maintenance for the application. This means that you might not see the maintenance configuration update applied to the maintenance process that follows a successful invocation of this operation, but to the following maintenance process instead.

To see the current maintenance configuration of your application, invoke the DescribeApplication operation.

For information about application maintenance, see Kinesis Data Analytics for Apache Flink Maintenance.

This operation is supported only for Amazon Kinesis Data Analytics for Apache Flink.

Creating a Request

data UpdateApplicationMaintenanceConfiguration

See: newUpdateApplicationMaintenanceConfiguration smart constructor.

Constructors

UpdateApplicationMaintenanceConfiguration'

Fields

applicationName :: Text

The name of the application for which you want to update the maintenance configuration.

applicationMaintenanceConfigurationUpdate :: ApplicationMaintenanceConfigurationUpdate

Describes the application maintenance configuration update.

newUpdateApplicationMaintenanceConfiguration

Arguments

:: Text
$sel:applicationName:UpdateApplicationMaintenanceConfiguration'

-> ApplicationMaintenanceConfigurationUpdate
$sel:applicationMaintenanceConfigurationUpdate:UpdateApplicationMaintenanceConfiguration'

-> UpdateApplicationMaintenanceConfiguration

Create a value of UpdateApplicationMaintenanceConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationName:UpdateApplicationMaintenanceConfiguration', updateApplicationMaintenanceConfiguration_applicationName - The name of the application for which you want to update the maintenance configuration.

$sel:applicationMaintenanceConfigurationUpdate:UpdateApplicationMaintenanceConfiguration', updateApplicationMaintenanceConfiguration_applicationMaintenanceConfigurationUpdate - Describes the application maintenance configuration update.

Request Lenses

updateApplicationMaintenanceConfiguration_applicationName :: Lens' UpdateApplicationMaintenanceConfiguration Text

The name of the application for which you want to update the maintenance configuration.

updateApplicationMaintenanceConfiguration_applicationMaintenanceConfigurationUpdate :: Lens' UpdateApplicationMaintenanceConfiguration ApplicationMaintenanceConfigurationUpdate

Describes the application maintenance configuration update.

Destructuring the Response

data UpdateApplicationMaintenanceConfigurationResponse

See: newUpdateApplicationMaintenanceConfigurationResponse smart constructor.

Constructors

UpdateApplicationMaintenanceConfigurationResponse'

Fields

applicationARN :: Maybe Text

The Amazon Resource Name (ARN) of the application.

applicationMaintenanceConfigurationDescription :: Maybe ApplicationMaintenanceConfigurationDescription

The application maintenance configuration description after the update.

httpStatus :: Int

The response's http status code.

newUpdateApplicationMaintenanceConfigurationResponse

Arguments

:: Int
$sel:httpStatus:UpdateApplicationMaintenanceConfigurationResponse'

-> UpdateApplicationMaintenanceConfigurationResponse

Create a value of UpdateApplicationMaintenanceConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationARN:UpdateApplicationMaintenanceConfigurationResponse', updateApplicationMaintenanceConfigurationResponse_applicationARN - The Amazon Resource Name (ARN) of the application.

$sel:applicationMaintenanceConfigurationDescription:UpdateApplicationMaintenanceConfigurationResponse', updateApplicationMaintenanceConfigurationResponse_applicationMaintenanceConfigurationDescription - The application maintenance configuration description after the update.

$sel:httpStatus:UpdateApplicationMaintenanceConfigurationResponse', updateApplicationMaintenanceConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

updateApplicationMaintenanceConfigurationResponse_applicationARN :: Lens' UpdateApplicationMaintenanceConfigurationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the application.

updateApplicationMaintenanceConfigurationResponse_applicationMaintenanceConfigurationDescription :: Lens' UpdateApplicationMaintenanceConfigurationResponse (Maybe ApplicationMaintenanceConfigurationDescription)

The application maintenance configuration description after the update.

updateApplicationMaintenanceConfigurationResponse_httpStatus :: Lens' UpdateApplicationMaintenanceConfigurationResponse Int

The response's http status code.