Network.AWS.KinesisVideo.CreateSignalingChannel

Description

Creates a signaling channel.

CreateSignalingChannel is an asynchronous operation.

Creating a Request

data CreateSignalingChannel

See: newCreateSignalingChannel smart constructor.

Constructors

CreateSignalingChannel'

Fields

singleMasterConfiguration :: Maybe SingleMasterConfiguration

A structure containing the configuration for the SINGLE_MASTER channel type.

channelType :: Maybe ChannelType

A type of the signaling channel that you are creating. Currently, SINGLE_MASTER is the only supported channel type.

tags :: Maybe [Tag]

A set of tags (key-value pairs) that you want to associate with this channel.

channelName :: Text

A name for the signaling channel that you are creating. It must be unique for each AWS account and AWS Region.

newCreateSignalingChannel

Arguments

:: Text
$sel:channelName:CreateSignalingChannel'

-> CreateSignalingChannel

Create a value of CreateSignalingChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:singleMasterConfiguration:CreateSignalingChannel', createSignalingChannel_singleMasterConfiguration - A structure containing the configuration for the SINGLE_MASTER channel type.

$sel:channelType:CreateSignalingChannel', createSignalingChannel_channelType - A type of the signaling channel that you are creating. Currently, SINGLE_MASTER is the only supported channel type.

$sel:tags:CreateSignalingChannel', createSignalingChannel_tags - A set of tags (key-value pairs) that you want to associate with this channel.

$sel:channelName:CreateSignalingChannel', createSignalingChannel_channelName - A name for the signaling channel that you are creating. It must be unique for each AWS account and AWS Region.

Request Lenses

createSignalingChannel_singleMasterConfiguration :: Lens' CreateSignalingChannel (Maybe SingleMasterConfiguration)

A structure containing the configuration for the SINGLE_MASTER channel type.

createSignalingChannel_channelType :: Lens' CreateSignalingChannel (Maybe ChannelType)

A type of the signaling channel that you are creating. Currently, SINGLE_MASTER is the only supported channel type.

createSignalingChannel_tags :: Lens' CreateSignalingChannel (Maybe [Tag])

A set of tags (key-value pairs) that you want to associate with this channel.

createSignalingChannel_channelName :: Lens' CreateSignalingChannel Text

A name for the signaling channel that you are creating. It must be unique for each AWS account and AWS Region.

Destructuring the Response

data CreateSignalingChannelResponse

See: newCreateSignalingChannelResponse smart constructor.

Constructors

CreateSignalingChannelResponse'

Fields

channelARN :: Maybe Text

The Amazon Resource Name (ARN) of the created channel.

httpStatus :: Int

The response's http status code.

newCreateSignalingChannelResponse

Arguments

:: Int
$sel:httpStatus:CreateSignalingChannelResponse'

-> CreateSignalingChannelResponse

Create a value of CreateSignalingChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelARN:CreateSignalingChannelResponse', createSignalingChannelResponse_channelARN - The Amazon Resource Name (ARN) of the created channel.

$sel:httpStatus:CreateSignalingChannelResponse', createSignalingChannelResponse_httpStatus - The response's http status code.

Response Lenses

createSignalingChannelResponse_channelARN :: Lens' CreateSignalingChannelResponse (Maybe Text)

The Amazon Resource Name (ARN) of the created channel.

createSignalingChannelResponse_httpStatus :: Lens' CreateSignalingChannelResponse Int

The response's http status code.