Network.AWS.KinesisVideo.CreateStream

Description

Creates a new Kinesis video stream.

When you create a new stream, Kinesis Video Streams assigns it a version number. When you change the stream's metadata, Kinesis Video Streams updates the version.

CreateStream is an asynchronous operation.

For information about how the service works, see How it Works.

You must have permissions for the KinesisVideo:CreateStream action.

Creating a Request

data CreateStream

See: newCreateStream smart constructor.

Constructors

CreateStream'

Fields

mediaType :: Maybe Text

The media type of the stream. Consumers of the stream can use this information when processing the stream. For more information about media types, see Media Types. If you choose to specify the MediaType, see Naming Requirements for guidelines.

Example valid values include "video/h264" and "video/h264,audio/aac".

This parameter is optional; the default value is null (or empty in JSON).

dataRetentionInHours :: Maybe Natural

The number of hours that you want to retain the data in the stream. Kinesis Video Streams retains the data in a data store that is associated with the stream.

The default value is 0, indicating that the stream does not persist data.

When the DataRetentionInHours value is 0, consumers can still consume the fragments that remain in the service host buffer, which has a retention time limit of 5 minutes and a retention memory limit of 200 MB. Fragments are removed from the buffer when either limit is reached.

kmsKeyId :: Maybe Text

The ID of the AWS Key Management Service (AWS KMS) key that you want Kinesis Video Streams to use to encrypt stream data.

If no key ID is specified, the default, Kinesis Video-managed key (aws/kinesisvideo) is used.

For more information, see DescribeKey.

deviceName :: Maybe Text

The name of the device that is writing to the stream.

In the current implementation, Kinesis Video Streams does not use this name.

tags :: Maybe (HashMap Text Text)

A list of tags to associate with the specified stream. Each tag is a key-value pair (the value is optional).

streamName :: Text

A name for the stream that you are creating.

The stream name is an identifier for the stream, and must be unique for each account and region.

newCreateStream

Arguments

:: Text
$sel:streamName:CreateStream'

-> CreateStream

Create a value of CreateStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mediaType:CreateStream', createStream_mediaType - The media type of the stream. Consumers of the stream can use this information when processing the stream. For more information about media types, see Media Types. If you choose to specify the MediaType, see Naming Requirements for guidelines.

Example valid values include "video/h264" and "video/h264,audio/aac".

This parameter is optional; the default value is null (or empty in JSON).

$sel:dataRetentionInHours:CreateStream', createStream_dataRetentionInHours - The number of hours that you want to retain the data in the stream. Kinesis Video Streams retains the data in a data store that is associated with the stream.

The default value is 0, indicating that the stream does not persist data.

When the DataRetentionInHours value is 0, consumers can still consume the fragments that remain in the service host buffer, which has a retention time limit of 5 minutes and a retention memory limit of 200 MB. Fragments are removed from the buffer when either limit is reached.

$sel:kmsKeyId:CreateStream', createStream_kmsKeyId - The ID of the AWS Key Management Service (AWS KMS) key that you want Kinesis Video Streams to use to encrypt stream data.

If no key ID is specified, the default, Kinesis Video-managed key (aws/kinesisvideo) is used.

For more information, see DescribeKey.

$sel:deviceName:CreateStream', createStream_deviceName - The name of the device that is writing to the stream.

In the current implementation, Kinesis Video Streams does not use this name.

$sel:tags:CreateStream', createStream_tags - A list of tags to associate with the specified stream. Each tag is a key-value pair (the value is optional).

$sel:streamName:CreateStream', createStream_streamName - A name for the stream that you are creating.

The stream name is an identifier for the stream, and must be unique for each account and region.

Request Lenses

createStream_mediaType :: Lens' CreateStream (Maybe Text)

The media type of the stream. Consumers of the stream can use this information when processing the stream. For more information about media types, see Media Types. If you choose to specify the MediaType, see Naming Requirements for guidelines.

Example valid values include "video/h264" and "video/h264,audio/aac".

This parameter is optional; the default value is null (or empty in JSON).

createStream_dataRetentionInHours :: Lens' CreateStream (Maybe Natural)

The number of hours that you want to retain the data in the stream. Kinesis Video Streams retains the data in a data store that is associated with the stream.

The default value is 0, indicating that the stream does not persist data.

When the DataRetentionInHours value is 0, consumers can still consume the fragments that remain in the service host buffer, which has a retention time limit of 5 minutes and a retention memory limit of 200 MB. Fragments are removed from the buffer when either limit is reached.

createStream_kmsKeyId :: Lens' CreateStream (Maybe Text)

The ID of the AWS Key Management Service (AWS KMS) key that you want Kinesis Video Streams to use to encrypt stream data.

If no key ID is specified, the default, Kinesis Video-managed key (aws/kinesisvideo) is used.

For more information, see DescribeKey.

createStream_deviceName :: Lens' CreateStream (Maybe Text)

The name of the device that is writing to the stream.

In the current implementation, Kinesis Video Streams does not use this name.

createStream_tags :: Lens' CreateStream (Maybe (HashMap Text Text))

A list of tags to associate with the specified stream. Each tag is a key-value pair (the value is optional).

createStream_streamName :: Lens' CreateStream Text

A name for the stream that you are creating.

The stream name is an identifier for the stream, and must be unique for each account and region.

Destructuring the Response

data CreateStreamResponse

See: newCreateStreamResponse smart constructor.

Constructors

CreateStreamResponse'

Fields

streamARN :: Maybe Text

The Amazon Resource Name (ARN) of the stream.

httpStatus :: Int

The response's http status code.

newCreateStreamResponse

Arguments

:: Int
$sel:httpStatus:CreateStreamResponse'

-> CreateStreamResponse

Create a value of CreateStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamARN:CreateStreamResponse', createStreamResponse_streamARN - The Amazon Resource Name (ARN) of the stream.

$sel:httpStatus:CreateStreamResponse', createStreamResponse_httpStatus - The response's http status code.

Response Lenses

createStreamResponse_streamARN :: Lens' CreateStreamResponse (Maybe Text)

The Amazon Resource Name (ARN) of the stream.

createStreamResponse_httpStatus :: Lens' CreateStreamResponse Int

The response's http status code.