Network.AWS.KinesisVideo.Lens

Description

Operations

ListTagsForStream

listTagsForStream_nextToken :: Lens' ListTagsForStream (Maybe Text)

If you specify this parameter and the result of a ListTagsForStream call is truncated, the response includes a token that you can use in the next request to fetch the next batch of tags.

listTagsForStream_streamARN :: Lens' ListTagsForStream (Maybe Text)

The Amazon Resource Name (ARN) of the stream that you want to list tags for.

listTagsForStream_streamName :: Lens' ListTagsForStream (Maybe Text)

The name of the stream that you want to list tags for.

listTagsForStreamResponse_nextToken :: Lens' ListTagsForStreamResponse (Maybe Text)

If you specify this parameter and the result of a ListTags call is truncated, the response includes a token that you can use in the next request to fetch the next set of tags.

listTagsForStreamResponse_tags :: Lens' ListTagsForStreamResponse (Maybe (HashMap Text Text))

A map of tag keys and values associated with the specified stream.

listTagsForStreamResponse_httpStatus :: Lens' ListTagsForStreamResponse Int

The response's http status code.

UntagResource

untagResource_resourceARN :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the signaling channel from which you want to remove tags.

untagResource_tagKeyList :: Lens' UntagResource (NonEmpty Text)

A list of the keys of the tags that you want to remove.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

TagResource

tagResource_resourceARN :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the signaling channel to which you want to add tags.

tagResource_tags :: Lens' TagResource (NonEmpty Tag)

A list of tags to associate with the specified signaling channel. Each tag is a key-value pair.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

DeleteStream

deleteStream_currentVersion :: Lens' DeleteStream (Maybe Text)

Optional: The version of the stream that you want to delete.

Specify the version as a safeguard to ensure that your are deleting the correct stream. To get the stream version, use the DescribeStream API.

If not specified, only the CreationTime is checked before deleting the stream.

deleteStream_streamARN :: Lens' DeleteStream Text

The Amazon Resource Name (ARN) of the stream that you want to delete.

deleteStreamResponse_httpStatus :: Lens' DeleteStreamResponse Int

The response's http status code.

UpdateStream

updateStream_deviceName :: Lens' UpdateStream (Maybe Text)

The name of the device that is writing to the stream.

In the current implementation, Kinesis Video Streams does not use this name.

updateStream_mediaType :: Lens' UpdateStream (Maybe Text)

The stream's media type. Use MediaType to specify the type of content that the stream contains to the consumers of the stream. For more information about media types, see Media Types. If you choose to specify the MediaType, see Naming Requirements.

To play video on the console, you must specify the correct video type. For example, if the video in the stream is H.264, specify video/h264 as the MediaType.

updateStream_streamARN :: Lens' UpdateStream (Maybe Text)

The ARN of the stream whose metadata you want to update.

updateStream_streamName :: Lens' UpdateStream (Maybe Text)

The name of the stream whose metadata you want to update.

The stream name is an identifier for the stream, and must be unique for each account and region.

updateStream_currentVersion :: Lens' UpdateStream Text

The version of the stream whose metadata you want to update.

updateStreamResponse_httpStatus :: Lens' UpdateStreamResponse Int

The response's http status code.

UpdateDataRetention

updateDataRetention_streamARN :: Lens' UpdateDataRetention (Maybe Text)

The Amazon Resource Name (ARN) of the stream whose retention period you want to change.

updateDataRetention_streamName :: Lens' UpdateDataRetention (Maybe Text)

The name of the stream whose retention period you want to change.

updateDataRetention_currentVersion :: Lens' UpdateDataRetention Text

The version of the stream whose retention period you want to change. To get the version, call either the DescribeStream or the ListStreams API.

updateDataRetention_operation :: Lens' UpdateDataRetention UpdateDataRetentionOperation

Indicates whether you want to increase or decrease the retention period.

updateDataRetention_dataRetentionChangeInHours :: Lens' UpdateDataRetention Natural

The retention period, in hours. The value you specify replaces the current value. The maximum value for this parameter is 87600 (ten years).

updateDataRetentionResponse_httpStatus :: Lens' UpdateDataRetentionResponse Int

The response's http status code.

GetDataEndpoint

getDataEndpoint_streamARN :: Lens' GetDataEndpoint (Maybe Text)

The Amazon Resource Name (ARN) of the stream that you want to get the endpoint for. You must specify either this parameter or a StreamName in the request.

getDataEndpoint_streamName :: Lens' GetDataEndpoint (Maybe Text)

The name of the stream that you want to get the endpoint for. You must specify either this parameter or a StreamARN in the request.

getDataEndpoint_aPIName :: Lens' GetDataEndpoint APIName

The name of the API action for which to get an endpoint.

getDataEndpointResponse_dataEndpoint :: Lens' GetDataEndpointResponse (Maybe Text)

The endpoint value. To read data from the stream or to write data to it, specify this endpoint in your application.

getDataEndpointResponse_httpStatus :: Lens' GetDataEndpointResponse Int

The response's http status code.

UntagStream

untagStream_streamARN :: Lens' UntagStream (Maybe Text)

The Amazon Resource Name (ARN) of the stream that you want to remove tags from.

untagStream_streamName :: Lens' UntagStream (Maybe Text)

The name of the stream that you want to remove tags from.

untagStream_tagKeyList :: Lens' UntagStream (NonEmpty Text)

A list of the keys of the tags that you want to remove.

untagStreamResponse_httpStatus :: Lens' UntagStreamResponse Int

The response's http status code.

DeleteSignalingChannel

deleteSignalingChannel_currentVersion :: Lens' DeleteSignalingChannel (Maybe Text)

The current version of the signaling channel that you want to delete. You can obtain the current version by invoking the DescribeSignalingChannel or ListSignalingChannels API operations.

deleteSignalingChannel_channelARN :: Lens' DeleteSignalingChannel Text

The Amazon Resource Name (ARN) of the signaling channel that you want to delete.

deleteSignalingChannelResponse_httpStatus :: Lens' DeleteSignalingChannelResponse Int

The response's http status code.

UpdateSignalingChannel

updateSignalingChannel_singleMasterConfiguration :: Lens' UpdateSignalingChannel (Maybe SingleMasterConfiguration)

The structure containing the configuration for the SINGLE_MASTER type of the signaling channel that you want to update.

updateSignalingChannel_channelARN :: Lens' UpdateSignalingChannel Text

The Amazon Resource Name (ARN) of the signaling channel that you want to update.

updateSignalingChannel_currentVersion :: Lens' UpdateSignalingChannel Text

The current version of the signaling channel that you want to update.

updateSignalingChannelResponse_httpStatus :: Lens' UpdateSignalingChannelResponse Int

The response's http status code.

ListSignalingChannels

listSignalingChannels_channelNameCondition :: Lens' ListSignalingChannels (Maybe ChannelNameCondition)

Optional: Returns only the channels that satisfy a specific condition.

listSignalingChannels_nextToken :: Lens' ListSignalingChannels (Maybe Text)

If you specify this parameter, when the result of a ListSignalingChannels operation is truncated, the call returns the NextToken in the response. To get another batch of channels, provide this token in your next request.

listSignalingChannels_maxResults :: Lens' ListSignalingChannels (Maybe Natural)

The maximum number of channels to return in the response. The default is 500.

listSignalingChannelsResponse_nextToken :: Lens' ListSignalingChannelsResponse (Maybe Text)

If the response is truncated, the call returns this element with a token. To get the next batch of streams, use this token in your next request.

listSignalingChannelsResponse_channelInfoList :: Lens' ListSignalingChannelsResponse (Maybe [ChannelInfo])

An array of ChannelInfo objects.

listSignalingChannelsResponse_httpStatus :: Lens' ListSignalingChannelsResponse Int

The response's http status code.

CreateSignalingChannel

createSignalingChannel_singleMasterConfiguration :: Lens' CreateSignalingChannel (Maybe SingleMasterConfiguration)

A structure containing the configuration for the SINGLE_MASTER channel type.

createSignalingChannel_channelType :: Lens' CreateSignalingChannel (Maybe ChannelType)

A type of the signaling channel that you are creating. Currently, SINGLE_MASTER is the only supported channel type.

createSignalingChannel_tags :: Lens' CreateSignalingChannel (Maybe [Tag])

A set of tags (key-value pairs) that you want to associate with this channel.

createSignalingChannel_channelName :: Lens' CreateSignalingChannel Text

A name for the signaling channel that you are creating. It must be unique for each AWS account and AWS Region.

createSignalingChannelResponse_channelARN :: Lens' CreateSignalingChannelResponse (Maybe Text)

The Amazon Resource Name (ARN) of the created channel.

createSignalingChannelResponse_httpStatus :: Lens' CreateSignalingChannelResponse Int

The response's http status code.

DescribeStream

describeStream_streamARN :: Lens' DescribeStream (Maybe Text)

The Amazon Resource Name (ARN) of the stream.

describeStream_streamName :: Lens' DescribeStream (Maybe Text)

The name of the stream.

describeStreamResponse_streamInfo :: Lens' DescribeStreamResponse (Maybe StreamInfo)

An object that describes the stream.

describeStreamResponse_httpStatus :: Lens' DescribeStreamResponse Int

The response's http status code.

TagStream

tagStream_streamARN :: Lens' TagStream (Maybe Text)

The Amazon Resource Name (ARN) of the resource that you want to add the tag or tags to.

tagStream_streamName :: Lens' TagStream (Maybe Text)

The name of the stream that you want to add the tag or tags to.

tagStream_tags :: Lens' TagStream (HashMap Text Text)

A list of tags to associate with the specified stream. Each tag is a key-value pair (the value is optional).

tagStreamResponse_httpStatus :: Lens' TagStreamResponse Int

The response's http status code.

DescribeSignalingChannel

describeSignalingChannel_channelName :: Lens' DescribeSignalingChannel (Maybe Text)

The name of the signaling channel that you want to describe.

describeSignalingChannel_channelARN :: Lens' DescribeSignalingChannel (Maybe Text)

The ARN of the signaling channel that you want to describe.

describeSignalingChannelResponse_channelInfo :: Lens' DescribeSignalingChannelResponse (Maybe ChannelInfo)

A structure that encapsulates the specified signaling channel's metadata and properties.

describeSignalingChannelResponse_httpStatus :: Lens' DescribeSignalingChannelResponse Int

The response's http status code.

GetSignalingChannelEndpoint

getSignalingChannelEndpoint_singleMasterChannelEndpointConfiguration :: Lens' GetSignalingChannelEndpoint (Maybe SingleMasterChannelEndpointConfiguration)

A structure containing the endpoint configuration for the SINGLE_MASTER channel type.

getSignalingChannelEndpoint_channelARN :: Lens' GetSignalingChannelEndpoint Text

The Amazon Resource Name (ARN) of the signalling channel for which you want to get an endpoint.

getSignalingChannelEndpointResponse_resourceEndpointList :: Lens' GetSignalingChannelEndpointResponse (Maybe [ResourceEndpointListItem])

A list of endpoints for the specified signaling channel.

getSignalingChannelEndpointResponse_httpStatus :: Lens' GetSignalingChannelEndpointResponse Int

The response's http status code.

CreateStream

createStream_dataRetentionInHours :: Lens' CreateStream (Maybe Natural)

The number of hours that you want to retain the data in the stream. Kinesis Video Streams retains the data in a data store that is associated with the stream.

The default value is 0, indicating that the stream does not persist data.

When the DataRetentionInHours value is 0, consumers can still consume the fragments that remain in the service host buffer, which has a retention time limit of 5 minutes and a retention memory limit of 200 MB. Fragments are removed from the buffer when either limit is reached.

createStream_kmsKeyId :: Lens' CreateStream (Maybe Text)

The ID of the AWS Key Management Service (AWS KMS) key that you want Kinesis Video Streams to use to encrypt stream data.

If no key ID is specified, the default, Kinesis Video-managed key (aws/kinesisvideo) is used.

For more information, see DescribeKey.

createStream_deviceName :: Lens' CreateStream (Maybe Text)

The name of the device that is writing to the stream.

In the current implementation, Kinesis Video Streams does not use this name.

createStream_mediaType :: Lens' CreateStream (Maybe Text)

The media type of the stream. Consumers of the stream can use this information when processing the stream. For more information about media types, see Media Types. If you choose to specify the MediaType, see Naming Requirements for guidelines.

Example valid values include "video/h264" and "video/h264,audio/aac".

This parameter is optional; the default value is null (or empty in JSON).

createStream_tags :: Lens' CreateStream (Maybe (HashMap Text Text))

A list of tags to associate with the specified stream. Each tag is a key-value pair (the value is optional).

createStream_streamName :: Lens' CreateStream Text

A name for the stream that you are creating.

The stream name is an identifier for the stream, and must be unique for each account and region.

createStreamResponse_streamARN :: Lens' CreateStreamResponse (Maybe Text)

The Amazon Resource Name (ARN) of the stream.

createStreamResponse_httpStatus :: Lens' CreateStreamResponse Int

The response's http status code.

ListStreams

listStreams_nextToken :: Lens' ListStreams (Maybe Text)

If you specify this parameter, when the result of a ListStreams operation is truncated, the call returns the NextToken in the response. To get another batch of streams, provide this token in your next request.

listStreams_maxResults :: Lens' ListStreams (Maybe Natural)

The maximum number of streams to return in the response. The default is 10,000.

listStreams_streamNameCondition :: Lens' ListStreams (Maybe StreamNameCondition)

Optional: Returns only streams that satisfy a specific condition. Currently, you can specify only the prefix of a stream name as a condition.

listStreamsResponse_nextToken :: Lens' ListStreamsResponse (Maybe Text)

If the response is truncated, the call returns this element with a token. To get the next batch of streams, use this token in your next request.

listStreamsResponse_streamInfoList :: Lens' ListStreamsResponse (Maybe [StreamInfo])

An array of StreamInfo objects.

listStreamsResponse_httpStatus :: Lens' ListStreamsResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_nextToken :: Lens' ListTagsForResource (Maybe Text)

If you specify this parameter and the result of a ListTagsForResource call is truncated, the response includes a token that you can use in the next request to fetch the next batch of tags.

listTagsForResource_resourceARN :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the signaling channel for which you want to list tags.

listTagsForResourceResponse_nextToken :: Lens' ListTagsForResourceResponse (Maybe Text)

If you specify this parameter and the result of a ListTagsForResource call is truncated, the response includes a token that you can use in the next request to fetch the next set of tags.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

A map of tag keys and values associated with the specified signaling channel.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

Types

ChannelInfo

channelInfo_creationTime :: Lens' ChannelInfo (Maybe UTCTime)

The time at which the signaling channel was created.

channelInfo_channelName :: Lens' ChannelInfo (Maybe Text)

The name of the signaling channel.

channelInfo_singleMasterConfiguration :: Lens' ChannelInfo (Maybe SingleMasterConfiguration)

A structure that contains the configuration for the SINGLE_MASTER channel type.

channelInfo_channelType :: Lens' ChannelInfo (Maybe ChannelType)

The type of the signaling channel.

channelInfo_version :: Lens' ChannelInfo (Maybe Text)

The current version of the signaling channel.

channelInfo_channelStatus :: Lens' ChannelInfo (Maybe StreamStatus)

Current status of the signaling channel.

channelInfo_channelARN :: Lens' ChannelInfo (Maybe Text)

The Amazon Resource Name (ARN) of the signaling channel.

ChannelNameCondition

channelNameCondition_comparisonOperator :: Lens' ChannelNameCondition (Maybe ComparisonOperator)

A comparison operator. Currently, you can only specify the BEGINS_WITH operator, which finds signaling channels whose names begin with a given prefix.

channelNameCondition_comparisonValue :: Lens' ChannelNameCondition (Maybe Text)

A value to compare.

ResourceEndpointListItem

resourceEndpointListItem_resourceEndpoint :: Lens' ResourceEndpointListItem (Maybe Text)

The endpoint of the signaling channel returned by the GetSignalingChannelEndpoint API.

resourceEndpointListItem_protocol :: Lens' ResourceEndpointListItem (Maybe ChannelProtocol)

The protocol of the signaling channel returned by the GetSignalingChannelEndpoint API.

SingleMasterChannelEndpointConfiguration

singleMasterChannelEndpointConfiguration_protocols :: Lens' SingleMasterChannelEndpointConfiguration (Maybe (NonEmpty ChannelProtocol))

This property is used to determine the nature of communication over this SINGLE_MASTER signaling channel. If WSS is specified, this API returns a websocket endpoint. If HTTPS is specified, this API returns an HTTPS endpoint.

singleMasterChannelEndpointConfiguration_role :: Lens' SingleMasterChannelEndpointConfiguration (Maybe ChannelRole)

This property is used to determine messaging permissions in this SINGLE_MASTER signaling channel. If MASTER is specified, this API returns an endpoint that a client can use to receive offers from and send answers to any of the viewers on this signaling channel. If VIEWER is specified, this API returns an endpoint that a client can use only to send offers to another MASTER client on this signaling channel.

SingleMasterConfiguration

singleMasterConfiguration_messageTtlSeconds :: Lens' SingleMasterConfiguration (Maybe Natural)

The period of time a signaling channel retains underlivered messages before they are discarded.

StreamInfo

streamInfo_creationTime :: Lens' StreamInfo (Maybe UTCTime)

A time stamp that indicates when the stream was created.

streamInfo_status :: Lens' StreamInfo (Maybe StreamStatus)

The status of the stream.

streamInfo_dataRetentionInHours :: Lens' StreamInfo (Maybe Natural)

How long the stream retains data, in hours.

streamInfo_version :: Lens' StreamInfo (Maybe Text)

The version of the stream.

streamInfo_kmsKeyId :: Lens' StreamInfo (Maybe Text)

The ID of the AWS Key Management Service (AWS KMS) key that Kinesis Video Streams uses to encrypt data on the stream.

streamInfo_deviceName :: Lens' StreamInfo (Maybe Text)

The name of the device that is associated with the stream.

streamInfo_mediaType :: Lens' StreamInfo (Maybe Text)

The MediaType of the stream.

streamInfo_streamARN :: Lens' StreamInfo (Maybe Text)

The Amazon Resource Name (ARN) of the stream.

streamInfo_streamName :: Lens' StreamInfo (Maybe Text)

The name of the stream.

StreamNameCondition

streamNameCondition_comparisonOperator :: Lens' StreamNameCondition (Maybe ComparisonOperator)

A comparison operator. Currently, you can specify only the BEGINS_WITH operator, which finds streams whose names start with a given prefix.

streamNameCondition_comparisonValue :: Lens' StreamNameCondition (Maybe Text)

A value to compare.

Tag

tag_key :: Lens' Tag Text

The key of the tag that is associated with the specified signaling channel.

tag_value :: Lens' Tag Text

The value of the tag that is associated with the specified signaling channel.