Network.AWS.KinesisVideo.ListSignalingChannels

Description

Returns an array of ChannelInfo objects. Each object describes a signaling channel. To retrieve only those channels that satisfy a specific condition, you can specify a ChannelNameCondition.

This operation returns paginated results.

Creating a Request

data ListSignalingChannels

See: newListSignalingChannels smart constructor.

Constructors

ListSignalingChannels'

Fields

channelNameCondition :: Maybe ChannelNameCondition

Optional: Returns only the channels that satisfy a specific condition.

nextToken :: Maybe Text

If you specify this parameter, when the result of a ListSignalingChannels operation is truncated, the call returns the NextToken in the response. To get another batch of channels, provide this token in your next request.

maxResults :: Maybe Natural

The maximum number of channels to return in the response. The default is 500.

newListSignalingChannels :: ListSignalingChannels

Create a value of ListSignalingChannels with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelNameCondition:ListSignalingChannels', listSignalingChannels_channelNameCondition - Optional: Returns only the channels that satisfy a specific condition.

$sel:nextToken:ListSignalingChannels', listSignalingChannels_nextToken - If you specify this parameter, when the result of a ListSignalingChannels operation is truncated, the call returns the NextToken in the response. To get another batch of channels, provide this token in your next request.

$sel:maxResults:ListSignalingChannels', listSignalingChannels_maxResults - The maximum number of channels to return in the response. The default is 500.

Request Lenses

listSignalingChannels_channelNameCondition :: Lens' ListSignalingChannels (Maybe ChannelNameCondition)

Optional: Returns only the channels that satisfy a specific condition.

listSignalingChannels_nextToken :: Lens' ListSignalingChannels (Maybe Text)

If you specify this parameter, when the result of a ListSignalingChannels operation is truncated, the call returns the NextToken in the response. To get another batch of channels, provide this token in your next request.

listSignalingChannels_maxResults :: Lens' ListSignalingChannels (Maybe Natural)

The maximum number of channels to return in the response. The default is 500.

Destructuring the Response

data ListSignalingChannelsResponse

See: newListSignalingChannelsResponse smart constructor.

Constructors

ListSignalingChannelsResponse'

Fields

channelInfoList :: Maybe [ChannelInfo]

An array of ChannelInfo objects.

nextToken :: Maybe Text

If the response is truncated, the call returns this element with a token. To get the next batch of streams, use this token in your next request.

httpStatus :: Int

The response's http status code.

newListSignalingChannelsResponse

Arguments

:: Int
$sel:httpStatus:ListSignalingChannelsResponse'

-> ListSignalingChannelsResponse

Create a value of ListSignalingChannelsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelInfoList:ListSignalingChannelsResponse', listSignalingChannelsResponse_channelInfoList - An array of ChannelInfo objects.

$sel:nextToken:ListSignalingChannels', listSignalingChannelsResponse_nextToken - If the response is truncated, the call returns this element with a token. To get the next batch of streams, use this token in your next request.

$sel:httpStatus:ListSignalingChannelsResponse', listSignalingChannelsResponse_httpStatus - The response's http status code.

Response Lenses

listSignalingChannelsResponse_channelInfoList :: Lens' ListSignalingChannelsResponse (Maybe [ChannelInfo])

An array of ChannelInfo objects.

listSignalingChannelsResponse_nextToken :: Lens' ListSignalingChannelsResponse (Maybe Text)

If the response is truncated, the call returns this element with a token. To get the next batch of streams, use this token in your next request.

listSignalingChannelsResponse_httpStatus :: Lens' ListSignalingChannelsResponse Int

The response's http status code.