Network.AWS.KinesisVideo.ListStreams

Description

Returns an array of StreamInfo objects. Each object describes a stream. To retrieve only streams that satisfy a specific condition, you can specify a StreamNameCondition.

This operation returns paginated results.

Creating a Request

data ListStreams

See: newListStreams smart constructor.

Constructors

ListStreams'

Fields

nextToken :: Maybe Text

If you specify this parameter, when the result of a ListStreams operation is truncated, the call returns the NextToken in the response. To get another batch of streams, provide this token in your next request.

streamNameCondition :: Maybe StreamNameCondition

Optional: Returns only streams that satisfy a specific condition. Currently, you can specify only the prefix of a stream name as a condition.

maxResults :: Maybe Natural

The maximum number of streams to return in the response. The default is 10,000.

newListStreams :: ListStreams

Create a value of ListStreams with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListStreams', listStreams_nextToken - If you specify this parameter, when the result of a ListStreams operation is truncated, the call returns the NextToken in the response. To get another batch of streams, provide this token in your next request.

$sel:streamNameCondition:ListStreams', listStreams_streamNameCondition - Optional: Returns only streams that satisfy a specific condition. Currently, you can specify only the prefix of a stream name as a condition.

$sel:maxResults:ListStreams', listStreams_maxResults - The maximum number of streams to return in the response. The default is 10,000.

Request Lenses

listStreams_nextToken :: Lens' ListStreams (Maybe Text)

If you specify this parameter, when the result of a ListStreams operation is truncated, the call returns the NextToken in the response. To get another batch of streams, provide this token in your next request.

listStreams_streamNameCondition :: Lens' ListStreams (Maybe StreamNameCondition)

Optional: Returns only streams that satisfy a specific condition. Currently, you can specify only the prefix of a stream name as a condition.

listStreams_maxResults :: Lens' ListStreams (Maybe Natural)

The maximum number of streams to return in the response. The default is 10,000.

Destructuring the Response

data ListStreamsResponse

See: newListStreamsResponse smart constructor.

Constructors

ListStreamsResponse'

Fields

streamInfoList :: Maybe [StreamInfo]

An array of StreamInfo objects.

nextToken :: Maybe Text

If the response is truncated, the call returns this element with a token. To get the next batch of streams, use this token in your next request.

httpStatus :: Int

The response's http status code.

newListStreamsResponse

Arguments

:: Int
$sel:httpStatus:ListStreamsResponse'

-> ListStreamsResponse

Create a value of ListStreamsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamInfoList:ListStreamsResponse', listStreamsResponse_streamInfoList - An array of StreamInfo objects.

$sel:nextToken:ListStreams', listStreamsResponse_nextToken - If the response is truncated, the call returns this element with a token. To get the next batch of streams, use this token in your next request.

$sel:httpStatus:ListStreamsResponse', listStreamsResponse_httpStatus - The response's http status code.

Response Lenses

listStreamsResponse_streamInfoList :: Lens' ListStreamsResponse (Maybe [StreamInfo])

An array of StreamInfo objects.

listStreamsResponse_nextToken :: Lens' ListStreamsResponse (Maybe Text)

If the response is truncated, the call returns this element with a token. To get the next batch of streams, use this token in your next request.

listStreamsResponse_httpStatus :: Lens' ListStreamsResponse Int

The response's http status code.