Network.AWS.KinesisVideo.Types

Description

Service Configuration

defaultService :: Service

API version 2017-09-30 of the Amazon Kinesis Video Streams SDK configuration.

Errors

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have required permissions to perform this operation.

_InvalidArgumentException :: AsError a => Getting (First ServiceError) a ServiceError

The value for this input parameter is invalid.

_TagsPerResourceExceededLimitException :: AsError a => Getting (First ServiceError) a ServiceError

You have exceeded the limit of tags that you can associate with the resource. Kinesis video streams support up to 50 tags.

_NotAuthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

The caller is not authorized to perform this operation.

_ClientLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Kinesis Video Streams has throttled the request because you have exceeded the limit of allowed client calls. Try making the call later.

_AccountChannelLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You have reached the maximum limit of active signaling channels for this AWS account in this region.

_InvalidDeviceException :: AsError a => Getting (First ServiceError) a ServiceError

Not implemented.

_VersionMismatchException :: AsError a => Getting (First ServiceError) a ServiceError

The stream version that you specified is not the latest version. To get the latest version, use the DescribeStream API.

_AccountStreamLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The number of streams created for the account is too high.

_InvalidResourceFormatException :: AsError a => Getting (First ServiceError) a ServiceError

The format of the StreamARN is invalid.

_DeviceStreamLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Not implemented.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Kinesis Video Streams can't find the stream that you specified.

_ResourceInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The signaling channel is currently not available for this operation.

APIName

newtype APIName

Constructors

APIName'

Fields

fromAPIName :: Text

Bundled Patterns

pattern APIName_GET_CLIP :: APIName

pattern APIName_GET_DASH_STREAMING_SESSION_URL :: APIName

pattern APIName_GET_HLS_STREAMING_SESSION_URL :: APIName

pattern APIName_GET_MEDIA :: APIName

pattern APIName_GET_MEDIA_FOR_FRAGMENT_LIST :: APIName

pattern APIName_LIST_FRAGMENTS :: APIName

pattern APIName_PUT_MEDIA :: APIName

ChannelProtocol

newtype ChannelProtocol

Constructors

ChannelProtocol'

Fields

fromChannelProtocol :: Text

Bundled Patterns

pattern ChannelProtocol_HTTPS :: ChannelProtocol

pattern ChannelProtocol_WSS :: ChannelProtocol

ChannelRole

newtype ChannelRole

Constructors

ChannelRole'

Fields

fromChannelRole :: Text

Bundled Patterns

pattern ChannelRole_MASTER :: ChannelRole

pattern ChannelRole_VIEWER :: ChannelRole

ChannelType

newtype ChannelType

Constructors

ChannelType'

Fields

fromChannelType :: Text

Bundled Patterns

pattern ChannelType_SINGLE_MASTER :: ChannelType

ComparisonOperator

newtype ComparisonOperator

Constructors

ComparisonOperator'

Fields

fromComparisonOperator :: Text

Bundled Patterns

pattern ComparisonOperator_BEGINS_WITH :: ComparisonOperator

StreamStatus

newtype StreamStatus

Constructors

StreamStatus'

Fields

fromStreamStatus :: Text

Bundled Patterns

pattern StreamStatus_ACTIVE :: StreamStatus

pattern StreamStatus_CREATING :: StreamStatus

pattern StreamStatus_DELETING :: StreamStatus

pattern StreamStatus_UPDATING :: StreamStatus

UpdateDataRetentionOperation

newtype UpdateDataRetentionOperation

Constructors

UpdateDataRetentionOperation'

Fields

fromUpdateDataRetentionOperation :: Text

Bundled Patterns

pattern UpdateDataRetentionOperation_DECREASE_DATA_RETENTION :: UpdateDataRetentionOperation

pattern UpdateDataRetentionOperation_INCREASE_DATA_RETENTION :: UpdateDataRetentionOperation

ChannelInfo

data ChannelInfo

A structure that encapsulates a signaling channel's metadata and properties.

See: newChannelInfo smart constructor.

Constructors

ChannelInfo'

Fields

creationTime :: Maybe POSIX

The time at which the signaling channel was created.

channelStatus :: Maybe StreamStatus

Current status of the signaling channel.

channelARN :: Maybe Text

The Amazon Resource Name (ARN) of the signaling channel.

singleMasterConfiguration :: Maybe SingleMasterConfiguration

A structure that contains the configuration for the SINGLE_MASTER channel type.

channelName :: Maybe Text

The name of the signaling channel.

version :: Maybe Text

The current version of the signaling channel.

channelType :: Maybe ChannelType

The type of the signaling channel.

newChannelInfo :: ChannelInfo

Create a value of ChannelInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ChannelInfo', channelInfo_creationTime - The time at which the signaling channel was created.

$sel:channelStatus:ChannelInfo', channelInfo_channelStatus - Current status of the signaling channel.

$sel:channelARN:ChannelInfo', channelInfo_channelARN - The Amazon Resource Name (ARN) of the signaling channel.

$sel:singleMasterConfiguration:ChannelInfo', channelInfo_singleMasterConfiguration - A structure that contains the configuration for the SINGLE_MASTER channel type.

$sel:channelName:ChannelInfo', channelInfo_channelName - The name of the signaling channel.

$sel:version:ChannelInfo', channelInfo_version - The current version of the signaling channel.

$sel:channelType:ChannelInfo', channelInfo_channelType - The type of the signaling channel.

channelInfo_creationTime :: Lens' ChannelInfo (Maybe UTCTime)

The time at which the signaling channel was created.

channelInfo_channelStatus :: Lens' ChannelInfo (Maybe StreamStatus)

Current status of the signaling channel.

channelInfo_channelARN :: Lens' ChannelInfo (Maybe Text)

The Amazon Resource Name (ARN) of the signaling channel.

channelInfo_singleMasterConfiguration :: Lens' ChannelInfo (Maybe SingleMasterConfiguration)

A structure that contains the configuration for the SINGLE_MASTER channel type.

channelInfo_channelName :: Lens' ChannelInfo (Maybe Text)

The name of the signaling channel.

channelInfo_version :: Lens' ChannelInfo (Maybe Text)

The current version of the signaling channel.

channelInfo_channelType :: Lens' ChannelInfo (Maybe ChannelType)

The type of the signaling channel.

ChannelNameCondition

data ChannelNameCondition

An optional input parameter for the ListSignalingChannels API. When this parameter is specified while invoking ListSignalingChannels, the API returns only the channels that satisfy a condition specified in ChannelNameCondition.

See: newChannelNameCondition smart constructor.

Constructors

ChannelNameCondition'

Fields

comparisonOperator :: Maybe ComparisonOperator

A comparison operator. Currently, you can only specify the BEGINS_WITH operator, which finds signaling channels whose names begin with a given prefix.

comparisonValue :: Maybe Text

A value to compare.

newChannelNameCondition :: ChannelNameCondition

Create a value of ChannelNameCondition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:comparisonOperator:ChannelNameCondition', channelNameCondition_comparisonOperator - A comparison operator. Currently, you can only specify the BEGINS_WITH operator, which finds signaling channels whose names begin with a given prefix.

$sel:comparisonValue:ChannelNameCondition', channelNameCondition_comparisonValue - A value to compare.

channelNameCondition_comparisonOperator :: Lens' ChannelNameCondition (Maybe ComparisonOperator)

A comparison operator. Currently, you can only specify the BEGINS_WITH operator, which finds signaling channels whose names begin with a given prefix.

channelNameCondition_comparisonValue :: Lens' ChannelNameCondition (Maybe Text)

A value to compare.

ResourceEndpointListItem

data ResourceEndpointListItem

An object that describes the endpoint of the signaling channel returned by the GetSignalingChannelEndpoint API.

See: newResourceEndpointListItem smart constructor.

Constructors

ResourceEndpointListItem'

Fields

protocol :: Maybe ChannelProtocol

The protocol of the signaling channel returned by the GetSignalingChannelEndpoint API.

resourceEndpoint :: Maybe Text

The endpoint of the signaling channel returned by the GetSignalingChannelEndpoint API.

newResourceEndpointListItem :: ResourceEndpointListItem

Create a value of ResourceEndpointListItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protocol:ResourceEndpointListItem', resourceEndpointListItem_protocol - The protocol of the signaling channel returned by the GetSignalingChannelEndpoint API.

$sel:resourceEndpoint:ResourceEndpointListItem', resourceEndpointListItem_resourceEndpoint - The endpoint of the signaling channel returned by the GetSignalingChannelEndpoint API.

resourceEndpointListItem_protocol :: Lens' ResourceEndpointListItem (Maybe ChannelProtocol)

The protocol of the signaling channel returned by the GetSignalingChannelEndpoint API.

resourceEndpointListItem_resourceEndpoint :: Lens' ResourceEndpointListItem (Maybe Text)

The endpoint of the signaling channel returned by the GetSignalingChannelEndpoint API.

SingleMasterChannelEndpointConfiguration

data SingleMasterChannelEndpointConfiguration

An object that contains the endpoint configuration for the SINGLE_MASTER channel type.

See: newSingleMasterChannelEndpointConfiguration smart constructor.

Constructors

SingleMasterChannelEndpointConfiguration'

Fields

protocols :: Maybe (NonEmpty ChannelProtocol)

This property is used to determine the nature of communication over this SINGLE_MASTER signaling channel. If WSS is specified, this API returns a websocket endpoint. If HTTPS is specified, this API returns an HTTPS endpoint.

role' :: Maybe ChannelRole

This property is used to determine messaging permissions in this SINGLE_MASTER signaling channel. If MASTER is specified, this API returns an endpoint that a client can use to receive offers from and send answers to any of the viewers on this signaling channel. If VIEWER is specified, this API returns an endpoint that a client can use only to send offers to another MASTER client on this signaling channel.

newSingleMasterChannelEndpointConfiguration :: SingleMasterChannelEndpointConfiguration

Create a value of SingleMasterChannelEndpointConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protocols:SingleMasterChannelEndpointConfiguration', singleMasterChannelEndpointConfiguration_protocols - This property is used to determine the nature of communication over this SINGLE_MASTER signaling channel. If WSS is specified, this API returns a websocket endpoint. If HTTPS is specified, this API returns an HTTPS endpoint.

$sel:role':SingleMasterChannelEndpointConfiguration', singleMasterChannelEndpointConfiguration_role - This property is used to determine messaging permissions in this SINGLE_MASTER signaling channel. If MASTER is specified, this API returns an endpoint that a client can use to receive offers from and send answers to any of the viewers on this signaling channel. If VIEWER is specified, this API returns an endpoint that a client can use only to send offers to another MASTER client on this signaling channel.

singleMasterChannelEndpointConfiguration_protocols :: Lens' SingleMasterChannelEndpointConfiguration (Maybe (NonEmpty ChannelProtocol))

This property is used to determine the nature of communication over this SINGLE_MASTER signaling channel. If WSS is specified, this API returns a websocket endpoint. If HTTPS is specified, this API returns an HTTPS endpoint.

singleMasterChannelEndpointConfiguration_role :: Lens' SingleMasterChannelEndpointConfiguration (Maybe ChannelRole)

This property is used to determine messaging permissions in this SINGLE_MASTER signaling channel. If MASTER is specified, this API returns an endpoint that a client can use to receive offers from and send answers to any of the viewers on this signaling channel. If VIEWER is specified, this API returns an endpoint that a client can use only to send offers to another MASTER client on this signaling channel.

SingleMasterConfiguration

data SingleMasterConfiguration

A structure that contains the configuration for the SINGLE_MASTER channel type.

See: newSingleMasterConfiguration smart constructor.

Constructors

SingleMasterConfiguration'

Fields

messageTtlSeconds :: Maybe Natural

The period of time a signaling channel retains underlivered messages before they are discarded.

newSingleMasterConfiguration :: SingleMasterConfiguration

Create a value of SingleMasterConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:messageTtlSeconds:SingleMasterConfiguration', singleMasterConfiguration_messageTtlSeconds - The period of time a signaling channel retains underlivered messages before they are discarded.

singleMasterConfiguration_messageTtlSeconds :: Lens' SingleMasterConfiguration (Maybe Natural)

The period of time a signaling channel retains underlivered messages before they are discarded.

StreamInfo

data StreamInfo

An object describing a Kinesis video stream.

See: newStreamInfo smart constructor.

Constructors

StreamInfo'

Fields

creationTime :: Maybe POSIX

A time stamp that indicates when the stream was created.

status :: Maybe StreamStatus

The status of the stream.

mediaType :: Maybe Text

The MediaType of the stream.

dataRetentionInHours :: Maybe Natural

How long the stream retains data, in hours.

streamARN :: Maybe Text

The Amazon Resource Name (ARN) of the stream.

kmsKeyId :: Maybe Text

The ID of the AWS Key Management Service (AWS KMS) key that Kinesis Video Streams uses to encrypt data on the stream.

deviceName :: Maybe Text

The name of the device that is associated with the stream.

version :: Maybe Text

The version of the stream.

streamName :: Maybe Text

The name of the stream.

newStreamInfo :: StreamInfo

Create a value of StreamInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:StreamInfo', streamInfo_creationTime - A time stamp that indicates when the stream was created.

$sel:status:StreamInfo', streamInfo_status - The status of the stream.

$sel:mediaType:StreamInfo', streamInfo_mediaType - The MediaType of the stream.

$sel:dataRetentionInHours:StreamInfo', streamInfo_dataRetentionInHours - How long the stream retains data, in hours.

$sel:streamARN:StreamInfo', streamInfo_streamARN - The Amazon Resource Name (ARN) of the stream.

$sel:kmsKeyId:StreamInfo', streamInfo_kmsKeyId - The ID of the AWS Key Management Service (AWS KMS) key that Kinesis Video Streams uses to encrypt data on the stream.

$sel:deviceName:StreamInfo', streamInfo_deviceName - The name of the device that is associated with the stream.

$sel:version:StreamInfo', streamInfo_version - The version of the stream.

$sel:streamName:StreamInfo', streamInfo_streamName - The name of the stream.

streamInfo_creationTime :: Lens' StreamInfo (Maybe UTCTime)

A time stamp that indicates when the stream was created.

streamInfo_status :: Lens' StreamInfo (Maybe StreamStatus)

The status of the stream.

streamInfo_mediaType :: Lens' StreamInfo (Maybe Text)

The MediaType of the stream.

streamInfo_dataRetentionInHours :: Lens' StreamInfo (Maybe Natural)

How long the stream retains data, in hours.

streamInfo_streamARN :: Lens' StreamInfo (Maybe Text)

The Amazon Resource Name (ARN) of the stream.

streamInfo_kmsKeyId :: Lens' StreamInfo (Maybe Text)

The ID of the AWS Key Management Service (AWS KMS) key that Kinesis Video Streams uses to encrypt data on the stream.

streamInfo_deviceName :: Lens' StreamInfo (Maybe Text)

The name of the device that is associated with the stream.

streamInfo_version :: Lens' StreamInfo (Maybe Text)

The version of the stream.

streamInfo_streamName :: Lens' StreamInfo (Maybe Text)

The name of the stream.

StreamNameCondition

data StreamNameCondition

Specifies the condition that streams must satisfy to be returned when you list streams (see the ListStreams API). A condition has a comparison operation and a value. Currently, you can specify only the BEGINS_WITH operator, which finds streams whose names start with a given prefix.

See: newStreamNameCondition smart constructor.

Constructors

StreamNameCondition'

Fields

comparisonOperator :: Maybe ComparisonOperator

A comparison operator. Currently, you can specify only the BEGINS_WITH operator, which finds streams whose names start with a given prefix.

comparisonValue :: Maybe Text

A value to compare.

newStreamNameCondition :: StreamNameCondition

Create a value of StreamNameCondition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:comparisonOperator:StreamNameCondition', streamNameCondition_comparisonOperator - A comparison operator. Currently, you can specify only the BEGINS_WITH operator, which finds streams whose names start with a given prefix.

$sel:comparisonValue:StreamNameCondition', streamNameCondition_comparisonValue - A value to compare.

streamNameCondition_comparisonOperator :: Lens' StreamNameCondition (Maybe ComparisonOperator)

A comparison operator. Currently, you can specify only the BEGINS_WITH operator, which finds streams whose names start with a given prefix.

streamNameCondition_comparisonValue :: Lens' StreamNameCondition (Maybe Text)

A value to compare.

Tag

data Tag

A key and value pair that is associated with the specified signaling channel.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Text

The key of the tag that is associated with the specified signaling channel.

value :: Text

The value of the tag that is associated with the specified signaling channel.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - The key of the tag that is associated with the specified signaling channel.

$sel:value:Tag', tag_value - The value of the tag that is associated with the specified signaling channel.

tag_key :: Lens' Tag Text

The key of the tag that is associated with the specified signaling channel.

tag_value :: Lens' Tag Text

The value of the tag that is associated with the specified signaling channel.