Network.AWS.KinesisVideo.UpdateStream

Description

Updates stream metadata, such as the device name and media type.

You must provide the stream name or the Amazon Resource Name (ARN) of the stream.

To make sure that you have the latest version of the stream before updating it, you can specify the stream version. Kinesis Video Streams assigns a version to each stream. When you update a stream, Kinesis Video Streams assigns a new version number. To get the latest stream version, use the DescribeStream API.

UpdateStream is an asynchronous operation, and takes time to complete.

Creating a Request

data UpdateStream

See: newUpdateStream smart constructor.

Constructors

UpdateStream'

Fields

deviceName :: Maybe Text

The name of the device that is writing to the stream.

In the current implementation, Kinesis Video Streams does not use this name.

mediaType :: Maybe Text

The stream's media type. Use MediaType to specify the type of content that the stream contains to the consumers of the stream. For more information about media types, see Media Types. If you choose to specify the MediaType, see Naming Requirements.

To play video on the console, you must specify the correct video type. For example, if the video in the stream is H.264, specify video/h264 as the MediaType.

streamARN :: Maybe Text

The ARN of the stream whose metadata you want to update.

streamName :: Maybe Text

The name of the stream whose metadata you want to update.

The stream name is an identifier for the stream, and must be unique for each account and region.

currentVersion :: Text

The version of the stream whose metadata you want to update.

newUpdateStream

Arguments

:: Text
$sel:currentVersion:UpdateStream'

-> UpdateStream

Create a value of UpdateStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceName:UpdateStream', updateStream_deviceName - The name of the device that is writing to the stream.

In the current implementation, Kinesis Video Streams does not use this name.

$sel:mediaType:UpdateStream', updateStream_mediaType - The stream's media type. Use MediaType to specify the type of content that the stream contains to the consumers of the stream. For more information about media types, see Media Types. If you choose to specify the MediaType, see Naming Requirements.

To play video on the console, you must specify the correct video type. For example, if the video in the stream is H.264, specify video/h264 as the MediaType.

$sel:streamARN:UpdateStream', updateStream_streamARN - The ARN of the stream whose metadata you want to update.

$sel:streamName:UpdateStream', updateStream_streamName - The name of the stream whose metadata you want to update.

The stream name is an identifier for the stream, and must be unique for each account and region.

$sel:currentVersion:UpdateStream', updateStream_currentVersion - The version of the stream whose metadata you want to update.

Request Lenses

updateStream_deviceName :: Lens' UpdateStream (Maybe Text)

The name of the device that is writing to the stream.

In the current implementation, Kinesis Video Streams does not use this name.

updateStream_mediaType :: Lens' UpdateStream (Maybe Text)

The stream's media type. Use MediaType to specify the type of content that the stream contains to the consumers of the stream. For more information about media types, see Media Types. If you choose to specify the MediaType, see Naming Requirements.

To play video on the console, you must specify the correct video type. For example, if the video in the stream is H.264, specify video/h264 as the MediaType.

updateStream_streamARN :: Lens' UpdateStream (Maybe Text)

The ARN of the stream whose metadata you want to update.

updateStream_streamName :: Lens' UpdateStream (Maybe Text)

The name of the stream whose metadata you want to update.

The stream name is an identifier for the stream, and must be unique for each account and region.

updateStream_currentVersion :: Lens' UpdateStream Text

The version of the stream whose metadata you want to update.

Destructuring the Response

data UpdateStreamResponse

See: newUpdateStreamResponse smart constructor.

Constructors

UpdateStreamResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateStreamResponse

Arguments

:: Int
$sel:httpStatus:UpdateStreamResponse'

-> UpdateStreamResponse

Create a value of UpdateStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateStreamResponse', updateStreamResponse_httpStatus - The response's http status code.

Response Lenses

updateStreamResponse_httpStatus :: Lens' UpdateStreamResponse Int

The response's http status code.