Network.AWS.KinesisVideoArchivedMedia

Description

Derived from API version 2017-09-30 of the AWS service descriptions, licensed under Apache 2.0.

Service Configuration

defaultService :: Service

API version 2017-09-30 of the Amazon Kinesis Video Streams Archived Media SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by KinesisVideoArchivedMedia.

InvalidMediaFrameException

_InvalidMediaFrameException :: AsError a => Getting (First ServiceError) a ServiceError

One or more frames in the requested clip could not be parsed based on the specified codec.

ClientLimitExceededException

_ClientLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Kinesis Video Streams has throttled the request because you have exceeded a limit. Try making the call later. For information about limits, see Kinesis Video Streams Limits.

InvalidCodecPrivateDataException

_InvalidCodecPrivateDataException :: AsError a => Getting (First ServiceError) a ServiceError

The codec private data in at least one of the tracks of the video stream is not valid for this operation.

NoDataRetentionException

_NoDataRetentionException :: AsError a => Getting (First ServiceError) a ServiceError

A streaming session was requested for a stream that does not retain data (that is, has a DataRetentionInHours of 0).

UnsupportedStreamMediaTypeException

_UnsupportedStreamMediaTypeException :: AsError a => Getting (First ServiceError) a ServiceError

The type of the media (for example, h.264 or h.265 video or ACC or G.711 audio) could not be determined from the codec IDs of the tracks in the first fragment for a playback session. The codec ID for track 1 should be V_MPEG/ISO/AVC and, optionally, the codec ID for track 2 should be A_AAC.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

GetMedia throws this error when Kinesis Video Streams can't find the stream that you specified.

GetHLSStreamingSessionURL and GetDASHStreamingSessionURL throw this error if a session with a PlaybackMode of ON_DEMAND or LIVE_REPLAYis requested for a stream that has no fragments within the requested time range, or if a session with a PlaybackMode of LIVE is requested for a stream that has no fragments within the last 30 seconds.

MissingCodecPrivateDataException

_MissingCodecPrivateDataException :: AsError a => Getting (First ServiceError) a ServiceError

No codec private data was found in at least one of tracks of the video stream.

NotAuthorizedException

_NotAuthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

Status Code: 403, The caller is not authorized to perform an operation on the given stream, or the token has expired.

InvalidArgumentException

_InvalidArgumentException :: AsError a => Getting (First ServiceError) a ServiceError

A specified parameter exceeds its restrictions, is not supported, or can't be used.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListFragments (Paginated)

data ListFragments

See: newListFragments smart constructor.

Constructors

ListFragments' (Maybe Text) (Maybe Natural) (Maybe Text) (Maybe Text) (Maybe FragmentSelector)

newListFragments :: ListFragments

Create a value of ListFragments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFragments', listFragments_nextToken - A token to specify where to start paginating. This is the ListFragmentsOutput$NextToken from a previously truncated response.

$sel:maxResults:ListFragments', listFragments_maxResults - The total number of fragments to return. If the total number of fragments available is more than the value specified in max-results, then a ListFragmentsOutput$NextToken is provided in the output that you can use to resume pagination.

$sel:streamARN:ListFragments', listFragments_streamARN - The Amazon Resource Name (ARN) of the stream from which to retrieve a fragment list. Specify either this parameter or the StreamName parameter.

$sel:streamName:ListFragments', listFragments_streamName - The name of the stream from which to retrieve a fragment list. Specify either this parameter or the StreamARN parameter.

$sel:fragmentSelector:ListFragments', listFragments_fragmentSelector - Describes the timestamp range and timestamp origin for the range of fragments to return.

data ListFragmentsResponse

See: newListFragmentsResponse smart constructor.

Constructors

ListFragmentsResponse' (Maybe Text) (Maybe [Fragment]) Int

newListFragmentsResponse

Arguments

:: Int
$sel:httpStatus:ListFragmentsResponse'

-> ListFragmentsResponse

Create a value of ListFragmentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFragments', listFragmentsResponse_nextToken - If the returned list is truncated, the operation returns this token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:fragments:ListFragmentsResponse', listFragmentsResponse_fragments - A list of archived Fragment objects from the stream that meet the selector criteria. Results are in no specific order, even across pages.

$sel:httpStatus:ListFragmentsResponse', listFragmentsResponse_httpStatus - The response's http status code.

GetMediaForFragmentList

data GetMediaForFragmentList

See: newGetMediaForFragmentList smart constructor.

Constructors

GetMediaForFragmentList' (Maybe Text) (Maybe Text) (NonEmpty Text)

newGetMediaForFragmentList

Arguments

:: NonEmpty Text
$sel:fragments:GetMediaForFragmentList'

-> GetMediaForFragmentList

Create a value of GetMediaForFragmentList with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamARN:GetMediaForFragmentList', getMediaForFragmentList_streamARN - The Amazon Resource Name (ARN) of the stream from which to retrieve fragment media. Specify either this parameter or the StreamName parameter.

$sel:streamName:GetMediaForFragmentList', getMediaForFragmentList_streamName - The name of the stream from which to retrieve fragment media. Specify either this parameter or the StreamARN parameter.

$sel:fragments:GetMediaForFragmentList', getMediaForFragmentList_fragments - A list of the numbers of fragments for which to retrieve media. You retrieve these values with ListFragments.

data GetMediaForFragmentListResponse

See: newGetMediaForFragmentListResponse smart constructor.

Constructors

GetMediaForFragmentListResponse' (Maybe Text) Int ResponseBody

newGetMediaForFragmentListResponse

Arguments

:: Int
$sel:httpStatus:GetMediaForFragmentListResponse'

-> ResponseBody
$sel:payload:GetMediaForFragmentListResponse'

-> GetMediaForFragmentListResponse

Create a value of GetMediaForFragmentListResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentType:GetMediaForFragmentListResponse', getMediaForFragmentListResponse_contentType - The content type of the requested media.

$sel:httpStatus:GetMediaForFragmentListResponse', getMediaForFragmentListResponse_httpStatus - The response's http status code.

$sel:payload:GetMediaForFragmentListResponse', getMediaForFragmentListResponse_payload - The payload that Kinesis Video Streams returns is a sequence of chunks from the specified stream. For information about the chunks, see PutMedia. The chunks that Kinesis Video Streams returns in the GetMediaForFragmentList call also include the following additional Matroska (MKV) tags:

AWS_KINESISVIDEO_FRAGMENT_NUMBER - Fragment number returned in the chunk.

AWS_KINESISVIDEO_SERVER_SIDE_TIMESTAMP - Server-side timestamp of the fragment.

AWS_KINESISVIDEO_PRODUCER_SIDE_TIMESTAMP - Producer-side timestamp of the fragment.

The following tags will be included if an exception occurs:

AWS_KINESISVIDEO_FRAGMENT_NUMBER - The number of the fragment that threw the exception

AWS_KINESISVIDEO_EXCEPTION_ERROR_CODE - The integer code of the exception

AWS_KINESISVIDEO_EXCEPTION_MESSAGE - A text description of the exception

GetClip

data GetClip

See: newGetClip smart constructor.

Constructors

GetClip' (Maybe Text) (Maybe Text) ClipFragmentSelector

newGetClip

Arguments

:: ClipFragmentSelector
$sel:clipFragmentSelector:GetClip'

-> GetClip

Create a value of GetClip with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamARN:GetClip', getClip_streamARN - The Amazon Resource Name (ARN) of the stream for which to retrieve the media clip.

You must specify either the StreamName or the StreamARN.

$sel:streamName:GetClip', getClip_streamName - The name of the stream for which to retrieve the media clip.

You must specify either the StreamName or the StreamARN.

$sel:clipFragmentSelector:GetClip', getClip_clipFragmentSelector - The time range of the requested clip and the source of the timestamps.

data GetClipResponse

See: newGetClipResponse smart constructor.

Constructors

GetClipResponse' (Maybe Text) Int ResponseBody

newGetClipResponse

Arguments

:: Int
$sel:httpStatus:GetClipResponse'

-> ResponseBody
$sel:payload:GetClipResponse'

-> GetClipResponse

Create a value of GetClipResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentType:GetClipResponse', getClipResponse_contentType - The content type of the media in the requested clip.

$sel:httpStatus:GetClipResponse', getClipResponse_httpStatus - The response's http status code.

$sel:payload:GetClipResponse', getClipResponse_payload - Traditional MP4 file that contains the media clip from the specified video stream. The output will contain the first 100 MB or the first 200 fragments from the specified start timestamp. For more information, see Kinesis Video Streams Limits.

GetDASHStreamingSessionURL

data GetDASHStreamingSessionURL

See: newGetDASHStreamingSessionURL smart constructor.

Constructors

GetDASHStreamingSessionURL' (Maybe DASHDisplayFragmentTimestamp) (Maybe DASHDisplayFragmentNumber) (Maybe Natural) (Maybe DASHFragmentSelector) (Maybe DASHPlaybackMode) (Maybe Text) (Maybe Text) (Maybe Natural)

newGetDASHStreamingSessionURL :: GetDASHStreamingSessionURL

Create a value of GetDASHStreamingSessionURL with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:displayFragmentTimestamp:GetDASHStreamingSessionURL', getDASHStreamingSessionURL_displayFragmentTimestamp - Per the MPEG-DASH specification, the wall-clock time of fragments in the manifest file can be derived using attributes in the manifest itself. However, typically, MPEG-DASH compatible media players do not properly handle gaps in the media timeline. Kinesis Video Streams adjusts the media timeline in the manifest file to enable playback of media with discontinuities. Therefore, the wall-clock time derived from the manifest file may be inaccurate. If DisplayFragmentTimestamp is set to ALWAYS, the accurate fragment timestamp is added to each S element in the manifest file with the attribute name “kvs:ts”. A custom MPEG-DASH media player is necessary to leverage this custom attribute.

The default value is NEVER. When DASHFragmentSelector is SERVER_TIMESTAMP, the timestamps will be the server start timestamps. Similarly, when DASHFragmentSelector is PRODUCER_TIMESTAMP, the timestamps will be the producer start timestamps.

$sel:displayFragmentNumber:GetDASHStreamingSessionURL', getDASHStreamingSessionURL_displayFragmentNumber - Fragments are identified in the manifest file based on their sequence number in the session. If DisplayFragmentNumber is set to ALWAYS, the Kinesis Video Streams fragment number is added to each S element in the manifest file with the attribute name “kvs:fn”. These fragment numbers can be used for logging or for use with other APIs (e.g. GetMedia and GetMediaForFragmentList). A custom MPEG-DASH media player is necessary to leverage these this custom attribute.

The default value is NEVER.

$sel:maxManifestFragmentResults:GetDASHStreamingSessionURL', getDASHStreamingSessionURL_maxManifestFragmentResults - The maximum number of fragments that are returned in the MPEG-DASH manifest.

When the PlaybackMode is LIVE, the most recent fragments are returned up to this value. When the PlaybackMode is ON_DEMAND, the oldest fragments are returned, up to this maximum number.

When there are a higher number of fragments available in a live MPEG-DASH manifest, video players often buffer content before starting playback. Increasing the buffer size increases the playback latency, but it decreases the likelihood that rebuffering will occur during playback. We recommend that a live MPEG-DASH manifest have a minimum of 3 fragments and a maximum of 10 fragments.

The default is 5 fragments if PlaybackMode is LIVE or LIVE_REPLAY, and 1,000 if PlaybackMode is ON_DEMAND.

The maximum value of 1,000 fragments corresponds to more than 16 minutes of video on streams with 1-second fragments, and more than 2 1/2 hours of video on streams with 10-second fragments.

$sel:dASHFragmentSelector:GetDASHStreamingSessionURL', getDASHStreamingSessionURL_dASHFragmentSelector - The time range of the requested fragment and the source of the timestamps.

This parameter is required if PlaybackMode is ON_DEMAND or LIVE_REPLAY. This parameter is optional if PlaybackMode is@ LIVE. If PlaybackMode is LIVE, the FragmentSelectorType can be set, but the TimestampRange should not be set. If PlaybackMode is ON_DEMAND or LIVE_REPLAY, both FragmentSelectorType and TimestampRange@ must be set.

$sel:playbackMode:GetDASHStreamingSessionURL', getDASHStreamingSessionURL_playbackMode - Whether to retrieve live, live replay, or archived, on-demand data.

Features of the three types of sessions include the following:

LIVE : For sessions of this type, the MPEG-DASH manifest is continually updated with the latest fragments as they become available. We recommend that the media player retrieve a new manifest on a one-second interval. When this type of session is played in a media player, the user interface typically displays a "live" notification, with no scrubber control for choosing the position in the playback window to display.

In LIVE mode, the newest available fragments are included in an MPEG-DASH manifest, even if there is a gap between fragments (that is, if a fragment is missing). A gap like this might cause a media player to halt or cause a jump in playback. In this mode, fragments are not added to the MPEG-DASH manifest if they are older than the newest fragment in the playlist. If the missing fragment becomes available after a subsequent fragment is added to the manifest, the older fragment is not added, and the gap is not filled.

LIVE_REPLAY : For sessions of this type, the MPEG-DASH manifest is updated similarly to how it is updated for LIVE mode except that it starts by including fragments from a given start time. Instead of fragments being added as they are ingested, fragments are added as the duration of the next fragment elapses. For example, if the fragments in the session are two seconds long, then a new fragment is added to the manifest every two seconds. This mode is useful to be able to start playback from when an event is detected and continue live streaming media that has not yet been ingested as of the time of the session creation. This mode is also useful to stream previously archived media without being limited by the 1,000 fragment limit in the ON_DEMAND mode.

ON_DEMAND : For sessions of this type, the MPEG-DASH manifest contains all the fragments for the session, up to the number that is specified in MaxManifestFragmentResults. The manifest must be retrieved only once for each session. When this type of session is played in a media player, the user interface typically displays a scrubber control for choosing the position in the playback window to display.

In all playback modes, if FragmentSelectorType is PRODUCER_TIMESTAMP, and if there are multiple fragments with the same start timestamp, the fragment that has the larger fragment number (that is, the newer fragment) is included in the MPEG-DASH manifest. The other fragments are not included. Fragments that have different timestamps but have overlapping durations are still included in the MPEG-DASH manifest. This can lead to unexpected behavior in the media player.

The default is LIVE.

$sel:streamARN:GetDASHStreamingSessionURL', getDASHStreamingSessionURL_streamARN - The Amazon Resource Name (ARN) of the stream for which to retrieve the MPEG-DASH manifest URL.

You must specify either the StreamName or the StreamARN.

$sel:streamName:GetDASHStreamingSessionURL', getDASHStreamingSessionURL_streamName - The name of the stream for which to retrieve the MPEG-DASH manifest URL.

You must specify either the StreamName or the StreamARN.

$sel:expires:GetDASHStreamingSessionURL', getDASHStreamingSessionURL_expires - The time in seconds until the requested session expires. This value can be between 300 (5 minutes) and 43200 (12 hours).

When a session expires, no new calls to GetDashManifest, GetMP4InitFragment, or GetMP4MediaFragment can be made for that session.

The default is 300 (5 minutes).

data GetDASHStreamingSessionURLResponse

See: newGetDASHStreamingSessionURLResponse smart constructor.

Constructors

GetDASHStreamingSessionURLResponse' (Maybe Text) Int

newGetDASHStreamingSessionURLResponse

Arguments

:: Int
$sel:httpStatus:GetDASHStreamingSessionURLResponse'

-> GetDASHStreamingSessionURLResponse

Create a value of GetDASHStreamingSessionURLResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dASHStreamingSessionURL:GetDASHStreamingSessionURLResponse', getDASHStreamingSessionURLResponse_dASHStreamingSessionURL - The URL (containing the session token) that a media player can use to retrieve the MPEG-DASH manifest.

$sel:httpStatus:GetDASHStreamingSessionURLResponse', getDASHStreamingSessionURLResponse_httpStatus - The response's http status code.

GetHLSStreamingSessionURL

data GetHLSStreamingSessionURL

See: newGetHLSStreamingSessionURL smart constructor.

Constructors

GetHLSStreamingSessionURL' (Maybe Natural) (Maybe ContainerFormat) (Maybe HLSDisplayFragmentTimestamp) (Maybe HLSPlaybackMode) (Maybe Text) (Maybe Text) (Maybe Natural) (Maybe HLSDiscontinuityMode) (Maybe HLSFragmentSelector)

newGetHLSStreamingSessionURL :: GetHLSStreamingSessionURL

Create a value of GetHLSStreamingSessionURL with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxMediaPlaylistFragmentResults:GetHLSStreamingSessionURL', getHLSStreamingSessionURL_maxMediaPlaylistFragmentResults - The maximum number of fragments that are returned in the HLS media playlists.

When the PlaybackMode is LIVE, the most recent fragments are returned up to this value. When the PlaybackMode is ON_DEMAND, the oldest fragments are returned, up to this maximum number.

When there are a higher number of fragments available in a live HLS media playlist, video players often buffer content before starting playback. Increasing the buffer size increases the playback latency, but it decreases the likelihood that rebuffering will occur during playback. We recommend that a live HLS media playlist have a minimum of 3 fragments and a maximum of 10 fragments.

The default is 5 fragments if PlaybackMode is LIVE or LIVE_REPLAY, and 1,000 if PlaybackMode is ON_DEMAND.

The maximum value of 5,000 fragments corresponds to more than 80 minutes of video on streams with 1-second fragments, and more than 13 hours of video on streams with 10-second fragments.

$sel:containerFormat:GetHLSStreamingSessionURL', getHLSStreamingSessionURL_containerFormat - Specifies which format should be used for packaging the media. Specifying the FRAGMENTED_MP4 container format packages the media into MP4 fragments (fMP4 or CMAF). This is the recommended packaging because there is minimal packaging overhead. The other container format option is MPEG_TS. HLS has supported MPEG TS chunks since it was released and is sometimes the only supported packaging on older HLS players. MPEG TS typically has a 5-25 percent packaging overhead. This means MPEG TS typically requires 5-25 percent more bandwidth and cost than fMP4.

The default is FRAGMENTED_MP4.

$sel:displayFragmentTimestamp:GetHLSStreamingSessionURL', getHLSStreamingSessionURL_displayFragmentTimestamp - Specifies when the fragment start timestamps should be included in the HLS media playlist. Typically, media players report the playhead position as a time relative to the start of the first fragment in the playback session. However, when the start timestamps are included in the HLS media playlist, some media players might report the current playhead as an absolute time based on the fragment timestamps. This can be useful for creating a playback experience that shows viewers the wall-clock time of the media.

The default is NEVER. When HLSFragmentSelector is SERVER_TIMESTAMP, the timestamps will be the server start timestamps. Similarly, when HLSFragmentSelector is PRODUCER_TIMESTAMP, the timestamps will be the producer start timestamps.

$sel:playbackMode:GetHLSStreamingSessionURL', getHLSStreamingSessionURL_playbackMode - Whether to retrieve live, live replay, or archived, on-demand data.

Features of the three types of sessions include the following:

LIVE : For sessions of this type, the HLS media playlist is continually updated with the latest fragments as they become available. We recommend that the media player retrieve a new playlist on a one-second interval. When this type of session is played in a media player, the user interface typically displays a "live" notification, with no scrubber control for choosing the position in the playback window to display.

In LIVE mode, the newest available fragments are included in an HLS media playlist, even if there is a gap between fragments (that is, if a fragment is missing). A gap like this might cause a media player to halt or cause a jump in playback. In this mode, fragments are not added to the HLS media playlist if they are older than the newest fragment in the playlist. If the missing fragment becomes available after a subsequent fragment is added to the playlist, the older fragment is not added, and the gap is not filled.

LIVE_REPLAY : For sessions of this type, the HLS media playlist is updated similarly to how it is updated for LIVE mode except that it starts by including fragments from a given start time. Instead of fragments being added as they are ingested, fragments are added as the duration of the next fragment elapses. For example, if the fragments in the session are two seconds long, then a new fragment is added to the media playlist every two seconds. This mode is useful to be able to start playback from when an event is detected and continue live streaming media that has not yet been ingested as of the time of the session creation. This mode is also useful to stream previously archived media without being limited by the 1,000 fragment limit in the ON_DEMAND mode.

ON_DEMAND : For sessions of this type, the HLS media playlist contains all the fragments for the session, up to the number that is specified in MaxMediaPlaylistFragmentResults. The playlist must be retrieved only once for each session. When this type of session is played in a media player, the user interface typically displays a scrubber control for choosing the position in the playback window to display.

In all playback modes, if FragmentSelectorType is PRODUCER_TIMESTAMP, and if there are multiple fragments with the same start timestamp, the fragment that has the largest fragment number (that is, the newest fragment) is included in the HLS media playlist. The other fragments are not included. Fragments that have different timestamps but have overlapping durations are still included in the HLS media playlist. This can lead to unexpected behavior in the media player.

The default is LIVE.

$sel:streamARN:GetHLSStreamingSessionURL', getHLSStreamingSessionURL_streamARN - The Amazon Resource Name (ARN) of the stream for which to retrieve the HLS master playlist URL.

You must specify either the StreamName or the StreamARN.

$sel:streamName:GetHLSStreamingSessionURL', getHLSStreamingSessionURL_streamName - The name of the stream for which to retrieve the HLS master playlist URL.

You must specify either the StreamName or the StreamARN.

$sel:expires:GetHLSStreamingSessionURL', getHLSStreamingSessionURL_expires - The time in seconds until the requested session expires. This value can be between 300 (5 minutes) and 43200 (12 hours).

When a session expires, no new calls to GetHLSMasterPlaylist, GetHLSMediaPlaylist, GetMP4InitFragment, GetMP4MediaFragment, or GetTSFragment can be made for that session.

The default is 300 (5 minutes).

$sel:discontinuityMode:GetHLSStreamingSessionURL', getHLSStreamingSessionURL_discontinuityMode - Specifies when flags marking discontinuities between fragments are added to the media playlists.

Media players typically build a timeline of media content to play, based on the timestamps of each fragment. This means that if there is any overlap or gap between fragments (as is typical if HLSFragmentSelector is set to SERVER_TIMESTAMP), the media player timeline will also have small gaps between fragments in some places, and will overwrite frames in other places. Gaps in the media player timeline can cause playback to stall and overlaps can cause playback to be jittery. When there are discontinuity flags between fragments, the media player is expected to reset the timeline, resulting in the next fragment being played immediately after the previous fragment.

The following modes are supported:

ALWAYS: a discontinuity marker is placed between every fragment in the HLS media playlist. It is recommended to use a value of ALWAYS if the fragment timestamps are not accurate.

NEVER: no discontinuity markers are placed anywhere. It is recommended to use a value of NEVER to ensure the media player timeline most accurately maps to the producer timestamps.

ON_DISCONTINUITY: a discontinuity marker is placed between fragments that have a gap or overlap of more than 50 milliseconds. For most playback scenarios, it is recommended to use a value of ON_DISCONTINUITY so that the media player timeline is only reset when there is a significant issue with the media timeline (e.g. a missing fragment).

The default is ALWAYS when HLSFragmentSelector is set to SERVER_TIMESTAMP, and NEVER when it is set to PRODUCER_TIMESTAMP.

$sel:hLSFragmentSelector:GetHLSStreamingSessionURL', getHLSStreamingSessionURL_hLSFragmentSelector - The time range of the requested fragment and the source of the timestamps.

This parameter is required if PlaybackMode is ON_DEMAND or LIVE_REPLAY. This parameter is optional if PlaybackMode is@ LIVE. If PlaybackMode is LIVE, the FragmentSelectorType can be set, but the TimestampRange should not be set. If PlaybackMode is ON_DEMAND or LIVE_REPLAY, both FragmentSelectorType and TimestampRange@ must be set.

data GetHLSStreamingSessionURLResponse

See: newGetHLSStreamingSessionURLResponse smart constructor.

Constructors

GetHLSStreamingSessionURLResponse' (Maybe Text) Int

newGetHLSStreamingSessionURLResponse

Arguments

:: Int
$sel:httpStatus:GetHLSStreamingSessionURLResponse'

-> GetHLSStreamingSessionURLResponse

Create a value of GetHLSStreamingSessionURLResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hLSStreamingSessionURL:GetHLSStreamingSessionURLResponse', getHLSStreamingSessionURLResponse_hLSStreamingSessionURL - The URL (containing the session token) that a media player can use to retrieve the HLS master playlist.

$sel:httpStatus:GetHLSStreamingSessionURLResponse', getHLSStreamingSessionURLResponse_httpStatus - The response's http status code.

Types

ClipFragmentSelectorType

newtype ClipFragmentSelectorType

Constructors

ClipFragmentSelectorType'

Fields

fromClipFragmentSelectorType :: Text

Bundled Patterns

pattern ClipFragmentSelectorType_PRODUCER_TIMESTAMP :: ClipFragmentSelectorType

pattern ClipFragmentSelectorType_SERVER_TIMESTAMP :: ClipFragmentSelectorType

ContainerFormat

newtype ContainerFormat

Constructors

ContainerFormat'

Fields

fromContainerFormat :: Text

Bundled Patterns

pattern ContainerFormat_FRAGMENTED_MP4 :: ContainerFormat

pattern ContainerFormat_MPEG_TS :: ContainerFormat

DASHDisplayFragmentNumber

newtype DASHDisplayFragmentNumber

Constructors

DASHDisplayFragmentNumber'

Fields

fromDASHDisplayFragmentNumber :: Text

Bundled Patterns

pattern DASHDisplayFragmentNumber_ALWAYS :: DASHDisplayFragmentNumber

pattern DASHDisplayFragmentNumber_NEVER :: DASHDisplayFragmentNumber

DASHDisplayFragmentTimestamp

newtype DASHDisplayFragmentTimestamp

Constructors

DASHDisplayFragmentTimestamp'

Fields

fromDASHDisplayFragmentTimestamp :: Text

Bundled Patterns

pattern DASHDisplayFragmentTimestamp_ALWAYS :: DASHDisplayFragmentTimestamp

pattern DASHDisplayFragmentTimestamp_NEVER :: DASHDisplayFragmentTimestamp

DASHFragmentSelectorType

newtype DASHFragmentSelectorType

Constructors

DASHFragmentSelectorType'

Fields

fromDASHFragmentSelectorType :: Text

Bundled Patterns

pattern DASHFragmentSelectorType_PRODUCER_TIMESTAMP :: DASHFragmentSelectorType

pattern DASHFragmentSelectorType_SERVER_TIMESTAMP :: DASHFragmentSelectorType

DASHPlaybackMode

newtype DASHPlaybackMode

Constructors

DASHPlaybackMode'

Fields

fromDASHPlaybackMode :: Text

Bundled Patterns

pattern DASHPlaybackMode_LIVE :: DASHPlaybackMode

pattern DASHPlaybackMode_LIVE_REPLAY :: DASHPlaybackMode

pattern DASHPlaybackMode_ON_DEMAND :: DASHPlaybackMode

FragmentSelectorType

newtype FragmentSelectorType

Constructors

FragmentSelectorType'

Fields

fromFragmentSelectorType :: Text

Bundled Patterns

pattern FragmentSelectorType_PRODUCER_TIMESTAMP :: FragmentSelectorType

pattern FragmentSelectorType_SERVER_TIMESTAMP :: FragmentSelectorType

HLSDiscontinuityMode

newtype HLSDiscontinuityMode

Constructors

HLSDiscontinuityMode'

Fields

fromHLSDiscontinuityMode :: Text

Bundled Patterns

pattern HLSDiscontinuityMode_ALWAYS :: HLSDiscontinuityMode

pattern HLSDiscontinuityMode_NEVER :: HLSDiscontinuityMode

pattern HLSDiscontinuityMode_ON_DISCONTINUITY :: HLSDiscontinuityMode

HLSDisplayFragmentTimestamp

newtype HLSDisplayFragmentTimestamp

Constructors

HLSDisplayFragmentTimestamp'

Fields

fromHLSDisplayFragmentTimestamp :: Text

Bundled Patterns

pattern HLSDisplayFragmentTimestamp_ALWAYS :: HLSDisplayFragmentTimestamp

pattern HLSDisplayFragmentTimestamp_NEVER :: HLSDisplayFragmentTimestamp

HLSFragmentSelectorType

newtype HLSFragmentSelectorType

Constructors

HLSFragmentSelectorType'

Fields

fromHLSFragmentSelectorType :: Text

Bundled Patterns

pattern HLSFragmentSelectorType_PRODUCER_TIMESTAMP :: HLSFragmentSelectorType

pattern HLSFragmentSelectorType_SERVER_TIMESTAMP :: HLSFragmentSelectorType

HLSPlaybackMode

newtype HLSPlaybackMode

Constructors

HLSPlaybackMode'

Fields

fromHLSPlaybackMode :: Text

Bundled Patterns

pattern HLSPlaybackMode_LIVE :: HLSPlaybackMode

pattern HLSPlaybackMode_LIVE_REPLAY :: HLSPlaybackMode

pattern HLSPlaybackMode_ON_DEMAND :: HLSPlaybackMode

ClipFragmentSelector

data ClipFragmentSelector

Describes the timestamp range and timestamp origin of a range of fragments.

Fragments that have duplicate producer timestamps are deduplicated. This means that if producers are producing a stream of fragments with producer timestamps that are approximately equal to the true clock time, the clip will contain all of the fragments within the requested timestamp range. If some fragments are ingested within the same time range and very different points in time, only the oldest ingested collection of fragments are returned.

See: newClipFragmentSelector smart constructor.

Constructors

ClipFragmentSelector' ClipFragmentSelectorType ClipTimestampRange

newClipFragmentSelector

Arguments

:: ClipFragmentSelectorType
$sel:fragmentSelectorType:ClipFragmentSelector'

-> ClipTimestampRange
$sel:timestampRange:ClipFragmentSelector'

-> ClipFragmentSelector

Create a value of ClipFragmentSelector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fragmentSelectorType:ClipFragmentSelector', clipFragmentSelector_fragmentSelectorType - The origin of the timestamps to use (Server or Producer).

$sel:timestampRange:ClipFragmentSelector', clipFragmentSelector_timestampRange - The range of timestamps to return.

ClipTimestampRange

data ClipTimestampRange

The range of timestamps for which to return fragments.

See: newClipTimestampRange smart constructor.

Constructors

ClipTimestampRange' POSIX POSIX

newClipTimestampRange

Arguments

:: UTCTime
$sel:startTimestamp:ClipTimestampRange'

-> UTCTime
$sel:endTimestamp:ClipTimestampRange'

-> ClipTimestampRange

Create a value of ClipTimestampRange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startTimestamp:ClipTimestampRange', clipTimestampRange_startTimestamp - The starting timestamp in the range of timestamps for which to return fragments.

Only fragments that start exactly at or after StartTimestamp are included in the session. Fragments that start before StartTimestamp and continue past it aren't included in the session. If FragmentSelectorType is SERVER_TIMESTAMP, the StartTimestamp must be later than the stream head.

$sel:endTimestamp:ClipTimestampRange', clipTimestampRange_endTimestamp - The end of the timestamp range for the requested media.

This value must be within 24 hours of the specified StartTimestamp, and it must be later than the StartTimestamp value. If FragmentSelectorType for the request is SERVER_TIMESTAMP, this value must be in the past.

This value is inclusive. The EndTimestamp is compared to the (starting) timestamp of the fragment. Fragments that start before the EndTimestamp value and continue past it are included in the session.

DASHFragmentSelector

data DASHFragmentSelector

Contains the range of timestamps for the requested media, and the source of the timestamps.

See: newDASHFragmentSelector smart constructor.

Constructors

DASHFragmentSelector' (Maybe DASHFragmentSelectorType) (Maybe DASHTimestampRange)

newDASHFragmentSelector :: DASHFragmentSelector

Create a value of DASHFragmentSelector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fragmentSelectorType:DASHFragmentSelector', dASHFragmentSelector_fragmentSelectorType - The source of the timestamps for the requested media.

When FragmentSelectorType is set to PRODUCER_TIMESTAMP and GetDASHStreamingSessionURLInput$PlaybackMode is ON_DEMAND or LIVE_REPLAY, the first fragment ingested with a producer timestamp within the specified FragmentSelector$TimestampRange is included in the media playlist. In addition, the fragments with producer timestamps within the TimestampRange ingested immediately following the first fragment (up to the GetDASHStreamingSessionURLInput$MaxManifestFragmentResults value) are included.

Fragments that have duplicate producer timestamps are deduplicated. This means that if producers are producing a stream of fragments with producer timestamps that are approximately equal to the true clock time, the MPEG-DASH manifest will contain all of the fragments within the requested timestamp range. If some fragments are ingested within the same time range and very different points in time, only the oldest ingested collection of fragments are returned.

When FragmentSelectorType is set to PRODUCER_TIMESTAMP and GetDASHStreamingSessionURLInput$PlaybackMode is LIVE, the producer timestamps are used in the MP4 fragments and for deduplication. But the most recently ingested fragments based on server timestamps are included in the MPEG-DASH manifest. This means that even if fragments ingested in the past have producer timestamps with values now, they are not included in the HLS media playlist.

The default is SERVER_TIMESTAMP.

$sel:timestampRange:DASHFragmentSelector', dASHFragmentSelector_timestampRange - The start and end of the timestamp range for the requested media.

This value should not be present if PlaybackType is LIVE.

DASHTimestampRange

data DASHTimestampRange

The start and end of the timestamp range for the requested media.

This value should not be present if PlaybackType is LIVE.

The values in DASHimestampRange are inclusive. Fragments that start exactly at or after the start time are included in the session. Fragments that start before the start time and continue past it are not included in the session.

See: newDASHTimestampRange smart constructor.

Constructors

DASHTimestampRange' (Maybe POSIX) (Maybe POSIX)

newDASHTimestampRange :: DASHTimestampRange

Create a value of DASHTimestampRange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endTimestamp:DASHTimestampRange', dASHTimestampRange_endTimestamp - The end of the timestamp range for the requested media. This value must be within 24 hours of the specified StartTimestamp, and it must be later than the StartTimestamp value.

If FragmentSelectorType for the request is SERVER_TIMESTAMP, this value must be in the past.

The EndTimestamp value is required for ON_DEMAND mode, but optional for LIVE_REPLAY mode. If the EndTimestamp is not set for LIVE_REPLAY mode then the session will continue to include newly ingested fragments until the session expires.

This value is inclusive. The EndTimestamp is compared to the (starting) timestamp of the fragment. Fragments that start before the EndTimestamp value and continue past it are included in the session.

$sel:startTimestamp:DASHTimestampRange', dASHTimestampRange_startTimestamp - The start of the timestamp range for the requested media.

If the DASHTimestampRange value is specified, the StartTimestamp value is required.

Only fragments that start exactly at or after StartTimestamp are included in the session. Fragments that start before StartTimestamp and continue past it aren't included in the session. If FragmentSelectorType is SERVER_TIMESTAMP, the StartTimestamp must be later than the stream head.

Fragment

data Fragment

Represents a segment of video or other time-delimited data.

See: newFragment smart constructor.

Constructors

Fragment' (Maybe POSIX) (Maybe POSIX) (Maybe Text) (Maybe Integer) (Maybe Integer)

newFragment :: Fragment

Create a value of Fragment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:producerTimestamp:Fragment', fragment_producerTimestamp - The timestamp from the producer corresponding to the fragment.

$sel:serverTimestamp:Fragment', fragment_serverTimestamp - The timestamp from the AWS server corresponding to the fragment.

$sel:fragmentNumber:Fragment', fragment_fragmentNumber - The unique identifier of the fragment. This value monotonically increases based on the ingestion order.

$sel:fragmentSizeInBytes:Fragment', fragment_fragmentSizeInBytes - The total fragment size, including information about the fragment and contained media data.

$sel:fragmentLengthInMilliseconds:Fragment', fragment_fragmentLengthInMilliseconds - The playback duration or other time value associated with the fragment.

FragmentSelector

data FragmentSelector

Describes the timestamp range and timestamp origin of a range of fragments.

Only fragments with a start timestamp greater than or equal to the given start time and less than or equal to the end time are returned. For example, if a stream contains fragments with the following start timestamps:

00:00:00

00:00:02

00:00:04

00:00:06

A fragment selector range with a start time of 00:00:01 and end time of 00:00:04 would return the fragments with start times of 00:00:02 and 00:00:04.

See: newFragmentSelector smart constructor.

Constructors

FragmentSelector' FragmentSelectorType TimestampRange

newFragmentSelector

Arguments

:: FragmentSelectorType
$sel:fragmentSelectorType:FragmentSelector'

-> TimestampRange
$sel:timestampRange:FragmentSelector'

-> FragmentSelector

Create a value of FragmentSelector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fragmentSelectorType:FragmentSelector', fragmentSelector_fragmentSelectorType - The origin of the timestamps to use (Server or Producer).

$sel:timestampRange:FragmentSelector', fragmentSelector_timestampRange - The range of timestamps to return.

HLSFragmentSelector

data HLSFragmentSelector

Contains the range of timestamps for the requested media, and the source of the timestamps.

See: newHLSFragmentSelector smart constructor.

Constructors

HLSFragmentSelector' (Maybe HLSFragmentSelectorType) (Maybe HLSTimestampRange)

newHLSFragmentSelector :: HLSFragmentSelector

Create a value of HLSFragmentSelector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fragmentSelectorType:HLSFragmentSelector', hLSFragmentSelector_fragmentSelectorType - The source of the timestamps for the requested media.

When FragmentSelectorType is set to PRODUCER_TIMESTAMP and GetHLSStreamingSessionURLInput$PlaybackMode is ON_DEMAND or LIVE_REPLAY, the first fragment ingested with a producer timestamp within the specified FragmentSelector$TimestampRange is included in the media playlist. In addition, the fragments with producer timestamps within the TimestampRange ingested immediately following the first fragment (up to the GetHLSStreamingSessionURLInput$MaxMediaPlaylistFragmentResults value) are included.

Fragments that have duplicate producer timestamps are deduplicated. This means that if producers are producing a stream of fragments with producer timestamps that are approximately equal to the true clock time, the HLS media playlists will contain all of the fragments within the requested timestamp range. If some fragments are ingested within the same time range and very different points in time, only the oldest ingested collection of fragments are returned.

When FragmentSelectorType is set to PRODUCER_TIMESTAMP and GetHLSStreamingSessionURLInput$PlaybackMode is LIVE, the producer timestamps are used in the MP4 fragments and for deduplication. But the most recently ingested fragments based on server timestamps are included in the HLS media playlist. This means that even if fragments ingested in the past have producer timestamps with values now, they are not included in the HLS media playlist.

The default is SERVER_TIMESTAMP.

$sel:timestampRange:HLSFragmentSelector', hLSFragmentSelector_timestampRange - The start and end of the timestamp range for the requested media.

This value should not be present if PlaybackType is LIVE.

HLSTimestampRange

data HLSTimestampRange

The start and end of the timestamp range for the requested media.

This value should not be present if PlaybackType is LIVE.

See: newHLSTimestampRange smart constructor.

Constructors

HLSTimestampRange' (Maybe POSIX) (Maybe POSIX)

newHLSTimestampRange :: HLSTimestampRange

Create a value of HLSTimestampRange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endTimestamp:HLSTimestampRange', hLSTimestampRange_endTimestamp - The end of the timestamp range for the requested media. This value must be within 24 hours of the specified StartTimestamp, and it must be later than the StartTimestamp value.

If FragmentSelectorType for the request is SERVER_TIMESTAMP, this value must be in the past.

The EndTimestamp value is required for ON_DEMAND mode, but optional for LIVE_REPLAY mode. If the EndTimestamp is not set for LIVE_REPLAY mode then the session will continue to include newly ingested fragments until the session expires.

This value is inclusive. The EndTimestamp is compared to the (starting) timestamp of the fragment. Fragments that start before the EndTimestamp value and continue past it are included in the session.

$sel:startTimestamp:HLSTimestampRange', hLSTimestampRange_startTimestamp - The start of the timestamp range for the requested media.

If the HLSTimestampRange value is specified, the StartTimestamp value is required.

Only fragments that start exactly at or after StartTimestamp are included in the session. Fragments that start before StartTimestamp and continue past it aren't included in the session. If FragmentSelectorType is SERVER_TIMESTAMP, the StartTimestamp must be later than the stream head.

TimestampRange

data TimestampRange

The range of timestamps for which to return fragments.

See: newTimestampRange smart constructor.

Constructors

TimestampRange' POSIX POSIX

newTimestampRange

Arguments

:: UTCTime
$sel:startTimestamp:TimestampRange'

-> UTCTime
$sel:endTimestamp:TimestampRange'

-> TimestampRange

Create a value of TimestampRange with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startTimestamp:TimestampRange', timestampRange_startTimestamp - The starting timestamp in the range of timestamps for which to return fragments.

$sel:endTimestamp:TimestampRange', timestampRange_endTimestamp - The ending timestamp in the range of timestamps for which to return fragments.