Network.AWS.KinesisVideoArchivedMedia.GetMediaForFragmentList

Description

Gets media for a list of fragments (specified by fragment number) from the archived data in an Amazon Kinesis video stream.

You must first call the GetDataEndpoint API to get an endpoint. Then send the GetMediaForFragmentList requests to this endpoint using the --endpoint-url parameter.

For limits, see Kinesis Video Streams Limits.

If an error is thrown after invoking a Kinesis Video Streams archived media API, in addition to the HTTP status code and the response body, it includes the following pieces of information:

x-amz-ErrorType HTTP header – contains a more specific error type in addition to what the HTTP status code provides.

x-amz-RequestId HTTP header – if you want to report an issue to AWS, the support team can better diagnose the problem if given the Request Id.

Both the HTTP status code and the ErrorType header can be utilized to make programmatic decisions about whether errors are retry-able and under what conditions, as well as provide information on what actions the client programmer might need to take in order to successfully try again.

For more information, see the Errors section at the bottom of this topic, as well as Common Errors.

Creating a Request

data GetMediaForFragmentList

See: newGetMediaForFragmentList smart constructor.

Constructors

GetMediaForFragmentList'

Fields

streamARN :: Maybe Text

The Amazon Resource Name (ARN) of the stream from which to retrieve fragment media. Specify either this parameter or the StreamName parameter.

streamName :: Maybe Text

The name of the stream from which to retrieve fragment media. Specify either this parameter or the StreamARN parameter.

fragments :: NonEmpty Text

A list of the numbers of fragments for which to retrieve media. You retrieve these values with ListFragments.

newGetMediaForFragmentList

Arguments

:: NonEmpty Text
$sel:fragments:GetMediaForFragmentList'

-> GetMediaForFragmentList

Create a value of GetMediaForFragmentList with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamARN:GetMediaForFragmentList', getMediaForFragmentList_streamARN - The Amazon Resource Name (ARN) of the stream from which to retrieve fragment media. Specify either this parameter or the StreamName parameter.

$sel:streamName:GetMediaForFragmentList', getMediaForFragmentList_streamName - The name of the stream from which to retrieve fragment media. Specify either this parameter or the StreamARN parameter.

$sel:fragments:GetMediaForFragmentList', getMediaForFragmentList_fragments - A list of the numbers of fragments for which to retrieve media. You retrieve these values with ListFragments.

Request Lenses

getMediaForFragmentList_streamARN :: Lens' GetMediaForFragmentList (Maybe Text)

The Amazon Resource Name (ARN) of the stream from which to retrieve fragment media. Specify either this parameter or the StreamName parameter.

getMediaForFragmentList_streamName :: Lens' GetMediaForFragmentList (Maybe Text)

The name of the stream from which to retrieve fragment media. Specify either this parameter or the StreamARN parameter.

getMediaForFragmentList_fragments :: Lens' GetMediaForFragmentList (NonEmpty Text)

A list of the numbers of fragments for which to retrieve media. You retrieve these values with ListFragments.

Destructuring the Response

data GetMediaForFragmentListResponse

See: newGetMediaForFragmentListResponse smart constructor.

Constructors

GetMediaForFragmentListResponse'

Fields

contentType :: Maybe Text

The content type of the requested media.

httpStatus :: Int

The response's http status code.

payload :: ResponseBody

The payload that Kinesis Video Streams returns is a sequence of chunks from the specified stream. For information about the chunks, see PutMedia. The chunks that Kinesis Video Streams returns in the GetMediaForFragmentList call also include the following additional Matroska (MKV) tags:

AWS_KINESISVIDEO_FRAGMENT_NUMBER - Fragment number returned in the chunk.

AWS_KINESISVIDEO_SERVER_SIDE_TIMESTAMP - Server-side timestamp of the fragment.

AWS_KINESISVIDEO_PRODUCER_SIDE_TIMESTAMP - Producer-side timestamp of the fragment.

The following tags will be included if an exception occurs:

AWS_KINESISVIDEO_FRAGMENT_NUMBER - The number of the fragment that threw the exception

AWS_KINESISVIDEO_EXCEPTION_ERROR_CODE - The integer code of the exception

AWS_KINESISVIDEO_EXCEPTION_MESSAGE - A text description of the exception

newGetMediaForFragmentListResponse

Arguments

:: Int
$sel:httpStatus:GetMediaForFragmentListResponse'

-> ResponseBody
$sel:payload:GetMediaForFragmentListResponse'

-> GetMediaForFragmentListResponse

Create a value of GetMediaForFragmentListResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contentType:GetMediaForFragmentListResponse', getMediaForFragmentListResponse_contentType - The content type of the requested media.

$sel:httpStatus:GetMediaForFragmentListResponse', getMediaForFragmentListResponse_httpStatus - The response's http status code.

$sel:payload:GetMediaForFragmentListResponse', getMediaForFragmentListResponse_payload - The payload that Kinesis Video Streams returns is a sequence of chunks from the specified stream. For information about the chunks, see PutMedia. The chunks that Kinesis Video Streams returns in the GetMediaForFragmentList call also include the following additional Matroska (MKV) tags:

AWS_KINESISVIDEO_FRAGMENT_NUMBER - Fragment number returned in the chunk.

AWS_KINESISVIDEO_SERVER_SIDE_TIMESTAMP - Server-side timestamp of the fragment.

AWS_KINESISVIDEO_PRODUCER_SIDE_TIMESTAMP - Producer-side timestamp of the fragment.

The following tags will be included if an exception occurs:

AWS_KINESISVIDEO_FRAGMENT_NUMBER - The number of the fragment that threw the exception

AWS_KINESISVIDEO_EXCEPTION_ERROR_CODE - The integer code of the exception

AWS_KINESISVIDEO_EXCEPTION_MESSAGE - A text description of the exception

Response Lenses

getMediaForFragmentListResponse_contentType :: Lens' GetMediaForFragmentListResponse (Maybe Text)

The content type of the requested media.

getMediaForFragmentListResponse_httpStatus :: Lens' GetMediaForFragmentListResponse Int

The response's http status code.

getMediaForFragmentListResponse_payload :: Lens' GetMediaForFragmentListResponse ResponseBody

The payload that Kinesis Video Streams returns is a sequence of chunks from the specified stream. For information about the chunks, see PutMedia. The chunks that Kinesis Video Streams returns in the GetMediaForFragmentList call also include the following additional Matroska (MKV) tags:

AWS_KINESISVIDEO_FRAGMENT_NUMBER - Fragment number returned in the chunk.

AWS_KINESISVIDEO_SERVER_SIDE_TIMESTAMP - Server-side timestamp of the fragment.

AWS_KINESISVIDEO_PRODUCER_SIDE_TIMESTAMP - Producer-side timestamp of the fragment.

The following tags will be included if an exception occurs:

AWS_KINESISVIDEO_FRAGMENT_NUMBER - The number of the fragment that threw the exception

AWS_KINESISVIDEO_EXCEPTION_ERROR_CODE - The integer code of the exception

AWS_KINESISVIDEO_EXCEPTION_MESSAGE - A text description of the exception