Network.AWS.KinesisVideoArchivedMedia.Lens

Description

Operations

ListFragments

listFragments_nextToken :: Lens' ListFragments (Maybe Text)

A token to specify where to start paginating. This is the ListFragmentsOutput$NextToken from a previously truncated response.

listFragments_maxResults :: Lens' ListFragments (Maybe Natural)

The total number of fragments to return. If the total number of fragments available is more than the value specified in max-results, then a ListFragmentsOutput$NextToken is provided in the output that you can use to resume pagination.

listFragments_streamARN :: Lens' ListFragments (Maybe Text)

The Amazon Resource Name (ARN) of the stream from which to retrieve a fragment list. Specify either this parameter or the StreamName parameter.

listFragments_streamName :: Lens' ListFragments (Maybe Text)

The name of the stream from which to retrieve a fragment list. Specify either this parameter or the StreamARN parameter.

listFragments_fragmentSelector :: Lens' ListFragments (Maybe FragmentSelector)

Describes the timestamp range and timestamp origin for the range of fragments to return.

listFragmentsResponse_nextToken :: Lens' ListFragmentsResponse (Maybe Text)

If the returned list is truncated, the operation returns this token to use to retrieve the next page of results. This value is null when there are no more results to return.

listFragmentsResponse_fragments :: Lens' ListFragmentsResponse (Maybe [Fragment])

A list of archived Fragment objects from the stream that meet the selector criteria. Results are in no specific order, even across pages.

listFragmentsResponse_httpStatus :: Lens' ListFragmentsResponse Int

The response's http status code.

GetMediaForFragmentList

getMediaForFragmentList_streamARN :: Lens' GetMediaForFragmentList (Maybe Text)

The Amazon Resource Name (ARN) of the stream from which to retrieve fragment media. Specify either this parameter or the StreamName parameter.

getMediaForFragmentList_streamName :: Lens' GetMediaForFragmentList (Maybe Text)

The name of the stream from which to retrieve fragment media. Specify either this parameter or the StreamARN parameter.

getMediaForFragmentList_fragments :: Lens' GetMediaForFragmentList (NonEmpty Text)

A list of the numbers of fragments for which to retrieve media. You retrieve these values with ListFragments.

getMediaForFragmentListResponse_contentType :: Lens' GetMediaForFragmentListResponse (Maybe Text)

The content type of the requested media.

getMediaForFragmentListResponse_httpStatus :: Lens' GetMediaForFragmentListResponse Int

The response's http status code.

getMediaForFragmentListResponse_payload :: Lens' GetMediaForFragmentListResponse ResponseBody

The payload that Kinesis Video Streams returns is a sequence of chunks from the specified stream. For information about the chunks, see PutMedia. The chunks that Kinesis Video Streams returns in the GetMediaForFragmentList call also include the following additional Matroska (MKV) tags:

AWS_KINESISVIDEO_FRAGMENT_NUMBER - Fragment number returned in the chunk.

AWS_KINESISVIDEO_SERVER_SIDE_TIMESTAMP - Server-side timestamp of the fragment.

AWS_KINESISVIDEO_PRODUCER_SIDE_TIMESTAMP - Producer-side timestamp of the fragment.

The following tags will be included if an exception occurs:

AWS_KINESISVIDEO_FRAGMENT_NUMBER - The number of the fragment that threw the exception

AWS_KINESISVIDEO_EXCEPTION_ERROR_CODE - The integer code of the exception

AWS_KINESISVIDEO_EXCEPTION_MESSAGE - A text description of the exception

GetClip

getClip_streamARN :: Lens' GetClip (Maybe Text)

The Amazon Resource Name (ARN) of the stream for which to retrieve the media clip.

You must specify either the StreamName or the StreamARN.

getClip_streamName :: Lens' GetClip (Maybe Text)

The name of the stream for which to retrieve the media clip.

You must specify either the StreamName or the StreamARN.

getClip_clipFragmentSelector :: Lens' GetClip ClipFragmentSelector

The time range of the requested clip and the source of the timestamps.

getClipResponse_contentType :: Lens' GetClipResponse (Maybe Text)

The content type of the media in the requested clip.

getClipResponse_httpStatus :: Lens' GetClipResponse Int

The response's http status code.

getClipResponse_payload :: Lens' GetClipResponse ResponseBody

Traditional MP4 file that contains the media clip from the specified video stream. The output will contain the first 100 MB or the first 200 fragments from the specified start timestamp. For more information, see Kinesis Video Streams Limits.

GetDASHStreamingSessionURL

getDASHStreamingSessionURL_displayFragmentTimestamp :: Lens' GetDASHStreamingSessionURL (Maybe DASHDisplayFragmentTimestamp)

Per the MPEG-DASH specification, the wall-clock time of fragments in the manifest file can be derived using attributes in the manifest itself. However, typically, MPEG-DASH compatible media players do not properly handle gaps in the media timeline. Kinesis Video Streams adjusts the media timeline in the manifest file to enable playback of media with discontinuities. Therefore, the wall-clock time derived from the manifest file may be inaccurate. If DisplayFragmentTimestamp is set to ALWAYS, the accurate fragment timestamp is added to each S element in the manifest file with the attribute name “kvs:ts”. A custom MPEG-DASH media player is necessary to leverage this custom attribute.

The default value is NEVER. When DASHFragmentSelector is SERVER_TIMESTAMP, the timestamps will be the server start timestamps. Similarly, when DASHFragmentSelector is PRODUCER_TIMESTAMP, the timestamps will be the producer start timestamps.

getDASHStreamingSessionURL_displayFragmentNumber :: Lens' GetDASHStreamingSessionURL (Maybe DASHDisplayFragmentNumber)

Fragments are identified in the manifest file based on their sequence number in the session. If DisplayFragmentNumber is set to ALWAYS, the Kinesis Video Streams fragment number is added to each S element in the manifest file with the attribute name “kvs:fn”. These fragment numbers can be used for logging or for use with other APIs (e.g. GetMedia and GetMediaForFragmentList). A custom MPEG-DASH media player is necessary to leverage these this custom attribute.

The default value is NEVER.

getDASHStreamingSessionURL_maxManifestFragmentResults :: Lens' GetDASHStreamingSessionURL (Maybe Natural)

The maximum number of fragments that are returned in the MPEG-DASH manifest.

When the PlaybackMode is LIVE, the most recent fragments are returned up to this value. When the PlaybackMode is ON_DEMAND, the oldest fragments are returned, up to this maximum number.

When there are a higher number of fragments available in a live MPEG-DASH manifest, video players often buffer content before starting playback. Increasing the buffer size increases the playback latency, but it decreases the likelihood that rebuffering will occur during playback. We recommend that a live MPEG-DASH manifest have a minimum of 3 fragments and a maximum of 10 fragments.

The default is 5 fragments if PlaybackMode is LIVE or LIVE_REPLAY, and 1,000 if PlaybackMode is ON_DEMAND.

The maximum value of 1,000 fragments corresponds to more than 16 minutes of video on streams with 1-second fragments, and more than 2 1/2 hours of video on streams with 10-second fragments.

getDASHStreamingSessionURL_dASHFragmentSelector :: Lens' GetDASHStreamingSessionURL (Maybe DASHFragmentSelector)

The time range of the requested fragment and the source of the timestamps.

This parameter is required if PlaybackMode is ON_DEMAND or LIVE_REPLAY. This parameter is optional if PlaybackMode is@ LIVE. If PlaybackMode is LIVE, the FragmentSelectorType can be set, but the TimestampRange should not be set. If PlaybackMode is ON_DEMAND or LIVE_REPLAY, both FragmentSelectorType and TimestampRange@ must be set.

getDASHStreamingSessionURL_playbackMode :: Lens' GetDASHStreamingSessionURL (Maybe DASHPlaybackMode)

Whether to retrieve live, live replay, or archived, on-demand data.

Features of the three types of sessions include the following:

LIVE : For sessions of this type, the MPEG-DASH manifest is continually updated with the latest fragments as they become available. We recommend that the media player retrieve a new manifest on a one-second interval. When this type of session is played in a media player, the user interface typically displays a "live" notification, with no scrubber control for choosing the position in the playback window to display.

In LIVE mode, the newest available fragments are included in an MPEG-DASH manifest, even if there is a gap between fragments (that is, if a fragment is missing). A gap like this might cause a media player to halt or cause a jump in playback. In this mode, fragments are not added to the MPEG-DASH manifest if they are older than the newest fragment in the playlist. If the missing fragment becomes available after a subsequent fragment is added to the manifest, the older fragment is not added, and the gap is not filled.

LIVE_REPLAY : For sessions of this type, the MPEG-DASH manifest is updated similarly to how it is updated for LIVE mode except that it starts by including fragments from a given start time. Instead of fragments being added as they are ingested, fragments are added as the duration of the next fragment elapses. For example, if the fragments in the session are two seconds long, then a new fragment is added to the manifest every two seconds. This mode is useful to be able to start playback from when an event is detected and continue live streaming media that has not yet been ingested as of the time of the session creation. This mode is also useful to stream previously archived media without being limited by the 1,000 fragment limit in the ON_DEMAND mode.

ON_DEMAND : For sessions of this type, the MPEG-DASH manifest contains all the fragments for the session, up to the number that is specified in MaxManifestFragmentResults. The manifest must be retrieved only once for each session. When this type of session is played in a media player, the user interface typically displays a scrubber control for choosing the position in the playback window to display.

In all playback modes, if FragmentSelectorType is PRODUCER_TIMESTAMP, and if there are multiple fragments with the same start timestamp, the fragment that has the larger fragment number (that is, the newer fragment) is included in the MPEG-DASH manifest. The other fragments are not included. Fragments that have different timestamps but have overlapping durations are still included in the MPEG-DASH manifest. This can lead to unexpected behavior in the media player.

The default is LIVE.

getDASHStreamingSessionURL_streamARN :: Lens' GetDASHStreamingSessionURL (Maybe Text)

The Amazon Resource Name (ARN) of the stream for which to retrieve the MPEG-DASH manifest URL.

You must specify either the StreamName or the StreamARN.

getDASHStreamingSessionURL_streamName :: Lens' GetDASHStreamingSessionURL (Maybe Text)

The name of the stream for which to retrieve the MPEG-DASH manifest URL.

You must specify either the StreamName or the StreamARN.

getDASHStreamingSessionURL_expires :: Lens' GetDASHStreamingSessionURL (Maybe Natural)

The time in seconds until the requested session expires. This value can be between 300 (5 minutes) and 43200 (12 hours).

When a session expires, no new calls to GetDashManifest, GetMP4InitFragment, or GetMP4MediaFragment can be made for that session.

The default is 300 (5 minutes).

getDASHStreamingSessionURLResponse_dASHStreamingSessionURL :: Lens' GetDASHStreamingSessionURLResponse (Maybe Text)

The URL (containing the session token) that a media player can use to retrieve the MPEG-DASH manifest.

getDASHStreamingSessionURLResponse_httpStatus :: Lens' GetDASHStreamingSessionURLResponse Int

The response's http status code.

GetHLSStreamingSessionURL

getHLSStreamingSessionURL_maxMediaPlaylistFragmentResults :: Lens' GetHLSStreamingSessionURL (Maybe Natural)

The maximum number of fragments that are returned in the HLS media playlists.

When the PlaybackMode is LIVE, the most recent fragments are returned up to this value. When the PlaybackMode is ON_DEMAND, the oldest fragments are returned, up to this maximum number.

When there are a higher number of fragments available in a live HLS media playlist, video players often buffer content before starting playback. Increasing the buffer size increases the playback latency, but it decreases the likelihood that rebuffering will occur during playback. We recommend that a live HLS media playlist have a minimum of 3 fragments and a maximum of 10 fragments.

The default is 5 fragments if PlaybackMode is LIVE or LIVE_REPLAY, and 1,000 if PlaybackMode is ON_DEMAND.

The maximum value of 5,000 fragments corresponds to more than 80 minutes of video on streams with 1-second fragments, and more than 13 hours of video on streams with 10-second fragments.

getHLSStreamingSessionURL_containerFormat :: Lens' GetHLSStreamingSessionURL (Maybe ContainerFormat)

Specifies which format should be used for packaging the media. Specifying the FRAGMENTED_MP4 container format packages the media into MP4 fragments (fMP4 or CMAF). This is the recommended packaging because there is minimal packaging overhead. The other container format option is MPEG_TS. HLS has supported MPEG TS chunks since it was released and is sometimes the only supported packaging on older HLS players. MPEG TS typically has a 5-25 percent packaging overhead. This means MPEG TS typically requires 5-25 percent more bandwidth and cost than fMP4.

The default is FRAGMENTED_MP4.

getHLSStreamingSessionURL_displayFragmentTimestamp :: Lens' GetHLSStreamingSessionURL (Maybe HLSDisplayFragmentTimestamp)

Specifies when the fragment start timestamps should be included in the HLS media playlist. Typically, media players report the playhead position as a time relative to the start of the first fragment in the playback session. However, when the start timestamps are included in the HLS media playlist, some media players might report the current playhead as an absolute time based on the fragment timestamps. This can be useful for creating a playback experience that shows viewers the wall-clock time of the media.

The default is NEVER. When HLSFragmentSelector is SERVER_TIMESTAMP, the timestamps will be the server start timestamps. Similarly, when HLSFragmentSelector is PRODUCER_TIMESTAMP, the timestamps will be the producer start timestamps.

getHLSStreamingSessionURL_playbackMode :: Lens' GetHLSStreamingSessionURL (Maybe HLSPlaybackMode)

Whether to retrieve live, live replay, or archived, on-demand data.

Features of the three types of sessions include the following:

LIVE : For sessions of this type, the HLS media playlist is continually updated with the latest fragments as they become available. We recommend that the media player retrieve a new playlist on a one-second interval. When this type of session is played in a media player, the user interface typically displays a "live" notification, with no scrubber control for choosing the position in the playback window to display.

In LIVE mode, the newest available fragments are included in an HLS media playlist, even if there is a gap between fragments (that is, if a fragment is missing). A gap like this might cause a media player to halt or cause a jump in playback. In this mode, fragments are not added to the HLS media playlist if they are older than the newest fragment in the playlist. If the missing fragment becomes available after a subsequent fragment is added to the playlist, the older fragment is not added, and the gap is not filled.

LIVE_REPLAY : For sessions of this type, the HLS media playlist is updated similarly to how it is updated for LIVE mode except that it starts by including fragments from a given start time. Instead of fragments being added as they are ingested, fragments are added as the duration of the next fragment elapses. For example, if the fragments in the session are two seconds long, then a new fragment is added to the media playlist every two seconds. This mode is useful to be able to start playback from when an event is detected and continue live streaming media that has not yet been ingested as of the time of the session creation. This mode is also useful to stream previously archived media without being limited by the 1,000 fragment limit in the ON_DEMAND mode.

ON_DEMAND : For sessions of this type, the HLS media playlist contains all the fragments for the session, up to the number that is specified in MaxMediaPlaylistFragmentResults. The playlist must be retrieved only once for each session. When this type of session is played in a media player, the user interface typically displays a scrubber control for choosing the position in the playback window to display.

In all playback modes, if FragmentSelectorType is PRODUCER_TIMESTAMP, and if there are multiple fragments with the same start timestamp, the fragment that has the largest fragment number (that is, the newest fragment) is included in the HLS media playlist. The other fragments are not included. Fragments that have different timestamps but have overlapping durations are still included in the HLS media playlist. This can lead to unexpected behavior in the media player.

The default is LIVE.

getHLSStreamingSessionURL_streamARN :: Lens' GetHLSStreamingSessionURL (Maybe Text)

The Amazon Resource Name (ARN) of the stream for which to retrieve the HLS master playlist URL.

You must specify either the StreamName or the StreamARN.

getHLSStreamingSessionURL_streamName :: Lens' GetHLSStreamingSessionURL (Maybe Text)

The name of the stream for which to retrieve the HLS master playlist URL.

You must specify either the StreamName or the StreamARN.

getHLSStreamingSessionURL_expires :: Lens' GetHLSStreamingSessionURL (Maybe Natural)

The time in seconds until the requested session expires. This value can be between 300 (5 minutes) and 43200 (12 hours).

When a session expires, no new calls to GetHLSMasterPlaylist, GetHLSMediaPlaylist, GetMP4InitFragment, GetMP4MediaFragment, or GetTSFragment can be made for that session.

The default is 300 (5 minutes).

getHLSStreamingSessionURL_discontinuityMode :: Lens' GetHLSStreamingSessionURL (Maybe HLSDiscontinuityMode)

Specifies when flags marking discontinuities between fragments are added to the media playlists.

Media players typically build a timeline of media content to play, based on the timestamps of each fragment. This means that if there is any overlap or gap between fragments (as is typical if HLSFragmentSelector is set to SERVER_TIMESTAMP), the media player timeline will also have small gaps between fragments in some places, and will overwrite frames in other places. Gaps in the media player timeline can cause playback to stall and overlaps can cause playback to be jittery. When there are discontinuity flags between fragments, the media player is expected to reset the timeline, resulting in the next fragment being played immediately after the previous fragment.

The following modes are supported:

ALWAYS: a discontinuity marker is placed between every fragment in the HLS media playlist. It is recommended to use a value of ALWAYS if the fragment timestamps are not accurate.

NEVER: no discontinuity markers are placed anywhere. It is recommended to use a value of NEVER to ensure the media player timeline most accurately maps to the producer timestamps.

ON_DISCONTINUITY: a discontinuity marker is placed between fragments that have a gap or overlap of more than 50 milliseconds. For most playback scenarios, it is recommended to use a value of ON_DISCONTINUITY so that the media player timeline is only reset when there is a significant issue with the media timeline (e.g. a missing fragment).

The default is ALWAYS when HLSFragmentSelector is set to SERVER_TIMESTAMP, and NEVER when it is set to PRODUCER_TIMESTAMP.

getHLSStreamingSessionURL_hLSFragmentSelector :: Lens' GetHLSStreamingSessionURL (Maybe HLSFragmentSelector)

The time range of the requested fragment and the source of the timestamps.

This parameter is required if PlaybackMode is ON_DEMAND or LIVE_REPLAY. This parameter is optional if PlaybackMode is@ LIVE. If PlaybackMode is LIVE, the FragmentSelectorType can be set, but the TimestampRange should not be set. If PlaybackMode is ON_DEMAND or LIVE_REPLAY, both FragmentSelectorType and TimestampRange@ must be set.

getHLSStreamingSessionURLResponse_hLSStreamingSessionURL :: Lens' GetHLSStreamingSessionURLResponse (Maybe Text)

The URL (containing the session token) that a media player can use to retrieve the HLS master playlist.

getHLSStreamingSessionURLResponse_httpStatus :: Lens' GetHLSStreamingSessionURLResponse Int

The response's http status code.

Types

ClipFragmentSelector

clipFragmentSelector_fragmentSelectorType :: Lens' ClipFragmentSelector ClipFragmentSelectorType

The origin of the timestamps to use (Server or Producer).

clipFragmentSelector_timestampRange :: Lens' ClipFragmentSelector ClipTimestampRange

The range of timestamps to return.

ClipTimestampRange

clipTimestampRange_startTimestamp :: Lens' ClipTimestampRange UTCTime

The starting timestamp in the range of timestamps for which to return fragments.

Only fragments that start exactly at or after StartTimestamp are included in the session. Fragments that start before StartTimestamp and continue past it aren't included in the session. If FragmentSelectorType is SERVER_TIMESTAMP, the StartTimestamp must be later than the stream head.

clipTimestampRange_endTimestamp :: Lens' ClipTimestampRange UTCTime

The end of the timestamp range for the requested media.

This value must be within 24 hours of the specified StartTimestamp, and it must be later than the StartTimestamp value. If FragmentSelectorType for the request is SERVER_TIMESTAMP, this value must be in the past.

This value is inclusive. The EndTimestamp is compared to the (starting) timestamp of the fragment. Fragments that start before the EndTimestamp value and continue past it are included in the session.

DASHFragmentSelector

dASHFragmentSelector_fragmentSelectorType :: Lens' DASHFragmentSelector (Maybe DASHFragmentSelectorType)

The source of the timestamps for the requested media.

When FragmentSelectorType is set to PRODUCER_TIMESTAMP and GetDASHStreamingSessionURLInput$PlaybackMode is ON_DEMAND or LIVE_REPLAY, the first fragment ingested with a producer timestamp within the specified FragmentSelector$TimestampRange is included in the media playlist. In addition, the fragments with producer timestamps within the TimestampRange ingested immediately following the first fragment (up to the GetDASHStreamingSessionURLInput$MaxManifestFragmentResults value) are included.

Fragments that have duplicate producer timestamps are deduplicated. This means that if producers are producing a stream of fragments with producer timestamps that are approximately equal to the true clock time, the MPEG-DASH manifest will contain all of the fragments within the requested timestamp range. If some fragments are ingested within the same time range and very different points in time, only the oldest ingested collection of fragments are returned.

When FragmentSelectorType is set to PRODUCER_TIMESTAMP and GetDASHStreamingSessionURLInput$PlaybackMode is LIVE, the producer timestamps are used in the MP4 fragments and for deduplication. But the most recently ingested fragments based on server timestamps are included in the MPEG-DASH manifest. This means that even if fragments ingested in the past have producer timestamps with values now, they are not included in the HLS media playlist.

The default is SERVER_TIMESTAMP.

dASHFragmentSelector_timestampRange :: Lens' DASHFragmentSelector (Maybe DASHTimestampRange)

The start and end of the timestamp range for the requested media.

This value should not be present if PlaybackType is LIVE.

DASHTimestampRange

dASHTimestampRange_endTimestamp :: Lens' DASHTimestampRange (Maybe UTCTime)

The end of the timestamp range for the requested media. This value must be within 24 hours of the specified StartTimestamp, and it must be later than the StartTimestamp value.

If FragmentSelectorType for the request is SERVER_TIMESTAMP, this value must be in the past.

The EndTimestamp value is required for ON_DEMAND mode, but optional for LIVE_REPLAY mode. If the EndTimestamp is not set for LIVE_REPLAY mode then the session will continue to include newly ingested fragments until the session expires.

This value is inclusive. The EndTimestamp is compared to the (starting) timestamp of the fragment. Fragments that start before the EndTimestamp value and continue past it are included in the session.

dASHTimestampRange_startTimestamp :: Lens' DASHTimestampRange (Maybe UTCTime)

The start of the timestamp range for the requested media.

If the DASHTimestampRange value is specified, the StartTimestamp value is required.

Only fragments that start exactly at or after StartTimestamp are included in the session. Fragments that start before StartTimestamp and continue past it aren't included in the session. If FragmentSelectorType is SERVER_TIMESTAMP, the StartTimestamp must be later than the stream head.

Fragment

fragment_producerTimestamp :: Lens' Fragment (Maybe UTCTime)

The timestamp from the producer corresponding to the fragment.

fragment_serverTimestamp :: Lens' Fragment (Maybe UTCTime)

The timestamp from the AWS server corresponding to the fragment.

fragment_fragmentNumber :: Lens' Fragment (Maybe Text)

The unique identifier of the fragment. This value monotonically increases based on the ingestion order.

fragment_fragmentSizeInBytes :: Lens' Fragment (Maybe Integer)

The total fragment size, including information about the fragment and contained media data.

fragment_fragmentLengthInMilliseconds :: Lens' Fragment (Maybe Integer)

The playback duration or other time value associated with the fragment.

FragmentSelector

fragmentSelector_fragmentSelectorType :: Lens' FragmentSelector FragmentSelectorType

The origin of the timestamps to use (Server or Producer).

fragmentSelector_timestampRange :: Lens' FragmentSelector TimestampRange

The range of timestamps to return.

HLSFragmentSelector

hLSFragmentSelector_fragmentSelectorType :: Lens' HLSFragmentSelector (Maybe HLSFragmentSelectorType)

The source of the timestamps for the requested media.

When FragmentSelectorType is set to PRODUCER_TIMESTAMP and GetHLSStreamingSessionURLInput$PlaybackMode is ON_DEMAND or LIVE_REPLAY, the first fragment ingested with a producer timestamp within the specified FragmentSelector$TimestampRange is included in the media playlist. In addition, the fragments with producer timestamps within the TimestampRange ingested immediately following the first fragment (up to the GetHLSStreamingSessionURLInput$MaxMediaPlaylistFragmentResults value) are included.

Fragments that have duplicate producer timestamps are deduplicated. This means that if producers are producing a stream of fragments with producer timestamps that are approximately equal to the true clock time, the HLS media playlists will contain all of the fragments within the requested timestamp range. If some fragments are ingested within the same time range and very different points in time, only the oldest ingested collection of fragments are returned.

When FragmentSelectorType is set to PRODUCER_TIMESTAMP and GetHLSStreamingSessionURLInput$PlaybackMode is LIVE, the producer timestamps are used in the MP4 fragments and for deduplication. But the most recently ingested fragments based on server timestamps are included in the HLS media playlist. This means that even if fragments ingested in the past have producer timestamps with values now, they are not included in the HLS media playlist.

The default is SERVER_TIMESTAMP.

hLSFragmentSelector_timestampRange :: Lens' HLSFragmentSelector (Maybe HLSTimestampRange)

The start and end of the timestamp range for the requested media.

This value should not be present if PlaybackType is LIVE.

HLSTimestampRange

hLSTimestampRange_endTimestamp :: Lens' HLSTimestampRange (Maybe UTCTime)

The end of the timestamp range for the requested media. This value must be within 24 hours of the specified StartTimestamp, and it must be later than the StartTimestamp value.

If FragmentSelectorType for the request is SERVER_TIMESTAMP, this value must be in the past.

The EndTimestamp value is required for ON_DEMAND mode, but optional for LIVE_REPLAY mode. If the EndTimestamp is not set for LIVE_REPLAY mode then the session will continue to include newly ingested fragments until the session expires.

This value is inclusive. The EndTimestamp is compared to the (starting) timestamp of the fragment. Fragments that start before the EndTimestamp value and continue past it are included in the session.

hLSTimestampRange_startTimestamp :: Lens' HLSTimestampRange (Maybe UTCTime)

The start of the timestamp range for the requested media.

If the HLSTimestampRange value is specified, the StartTimestamp value is required.

Only fragments that start exactly at or after StartTimestamp are included in the session. Fragments that start before StartTimestamp and continue past it aren't included in the session. If FragmentSelectorType is SERVER_TIMESTAMP, the StartTimestamp must be later than the stream head.

TimestampRange

timestampRange_startTimestamp :: Lens' TimestampRange UTCTime

The starting timestamp in the range of timestamps for which to return fragments.

timestampRange_endTimestamp :: Lens' TimestampRange UTCTime

The ending timestamp in the range of timestamps for which to return fragments.