Network.AWS.KinesisVideoArchivedMedia.Types.ClipFragmentSelector

Description

Documentation

data ClipFragmentSelector

Describes the timestamp range and timestamp origin of a range of fragments.

Fragments that have duplicate producer timestamps are deduplicated. This means that if producers are producing a stream of fragments with producer timestamps that are approximately equal to the true clock time, the clip will contain all of the fragments within the requested timestamp range. If some fragments are ingested within the same time range and very different points in time, only the oldest ingested collection of fragments are returned.

See: newClipFragmentSelector smart constructor.

Constructors

ClipFragmentSelector'

Fields

fragmentSelectorType :: ClipFragmentSelectorType

The origin of the timestamps to use (Server or Producer).

timestampRange :: ClipTimestampRange

The range of timestamps to return.

newClipFragmentSelector

Arguments

:: ClipFragmentSelectorType
$sel:fragmentSelectorType:ClipFragmentSelector'

-> ClipTimestampRange
$sel:timestampRange:ClipFragmentSelector'

-> ClipFragmentSelector

Create a value of ClipFragmentSelector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fragmentSelectorType:ClipFragmentSelector', clipFragmentSelector_fragmentSelectorType - The origin of the timestamps to use (Server or Producer).

$sel:timestampRange:ClipFragmentSelector', clipFragmentSelector_timestampRange - The range of timestamps to return.

clipFragmentSelector_fragmentSelectorType :: Lens' ClipFragmentSelector ClipFragmentSelectorType

The origin of the timestamps to use (Server or Producer).

clipFragmentSelector_timestampRange :: Lens' ClipFragmentSelector ClipTimestampRange

The range of timestamps to return.