Network.AWS.KinesisVideoArchivedMedia.Types.Fragment

Description

Documentation

data Fragment

Represents a segment of video or other time-delimited data.

See: newFragment smart constructor.

Constructors

Fragment'

Fields

producerTimestamp :: Maybe POSIX

The timestamp from the producer corresponding to the fragment.

serverTimestamp :: Maybe POSIX

The timestamp from the AWS server corresponding to the fragment.

fragmentNumber :: Maybe Text

The unique identifier of the fragment. This value monotonically increases based on the ingestion order.

fragmentSizeInBytes :: Maybe Integer

The total fragment size, including information about the fragment and contained media data.

fragmentLengthInMilliseconds :: Maybe Integer

The playback duration or other time value associated with the fragment.

newFragment :: Fragment

Create a value of Fragment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:producerTimestamp:Fragment', fragment_producerTimestamp - The timestamp from the producer corresponding to the fragment.

$sel:serverTimestamp:Fragment', fragment_serverTimestamp - The timestamp from the AWS server corresponding to the fragment.

$sel:fragmentNumber:Fragment', fragment_fragmentNumber - The unique identifier of the fragment. This value monotonically increases based on the ingestion order.

$sel:fragmentSizeInBytes:Fragment', fragment_fragmentSizeInBytes - The total fragment size, including information about the fragment and contained media data.

$sel:fragmentLengthInMilliseconds:Fragment', fragment_fragmentLengthInMilliseconds - The playback duration or other time value associated with the fragment.

fragment_producerTimestamp :: Lens' Fragment (Maybe UTCTime)

The timestamp from the producer corresponding to the fragment.

fragment_serverTimestamp :: Lens' Fragment (Maybe UTCTime)

The timestamp from the AWS server corresponding to the fragment.

fragment_fragmentNumber :: Lens' Fragment (Maybe Text)

The unique identifier of the fragment. This value monotonically increases based on the ingestion order.

fragment_fragmentSizeInBytes :: Lens' Fragment (Maybe Integer)

The total fragment size, including information about the fragment and contained media data.

fragment_fragmentLengthInMilliseconds :: Lens' Fragment (Maybe Integer)

The playback duration or other time value associated with the fragment.