Network.AWS.KinesisVideoSignaling.Lens

Description

Operations

SendAlexaOfferToMaster

sendAlexaOfferToMaster_channelARN :: Lens' SendAlexaOfferToMaster Text

The ARN of the signaling channel by which Alexa and the master peer communicate.

sendAlexaOfferToMaster_senderClientId :: Lens' SendAlexaOfferToMaster Text

The unique identifier for the sender client.

sendAlexaOfferToMaster_messagePayload :: Lens' SendAlexaOfferToMaster Text

The base64-encoded SDP offer content.

sendAlexaOfferToMasterResponse_answer :: Lens' SendAlexaOfferToMasterResponse (Maybe Text)

The base64-encoded SDP answer content.

sendAlexaOfferToMasterResponse_httpStatus :: Lens' SendAlexaOfferToMasterResponse Int

The response's http status code.

GetIceServerConfig

getIceServerConfig_clientId :: Lens' GetIceServerConfig (Maybe Text)

Unique identifier for the viewer. Must be unique within the signaling channel.

getIceServerConfig_service :: Lens' GetIceServerConfig (Maybe Service)

Specifies the desired service. Currently, TURN is the only valid value.

getIceServerConfig_username :: Lens' GetIceServerConfig (Maybe Text)

An optional user ID to be associated with the credentials.

getIceServerConfig_channelARN :: Lens' GetIceServerConfig Text

The ARN of the signaling channel to be used for the peer-to-peer connection between configured peers.

getIceServerConfigResponse_iceServerList :: Lens' GetIceServerConfigResponse (Maybe [IceServer])

The list of ICE server information objects.

getIceServerConfigResponse_httpStatus :: Lens' GetIceServerConfigResponse Int

The response's http status code.

Types

IceServer

iceServer_ttl :: Lens' IceServer (Maybe Natural)

The period of time, in seconds, during which the username and password are valid.

iceServer_uris :: Lens' IceServer (Maybe [Text])

An array of URIs, in the form specified in the I-D.petithuguenin-behave-turn-uris spec. These URIs provide the different addresses and/or protocols that can be used to reach the TURN server.

iceServer_username :: Lens' IceServer (Maybe Text)

A username to login to the ICE server.

iceServer_password :: Lens' IceServer (Maybe Text)

A password to login to the ICE server.