Network.AWS.LakeFormation.AddLFTagsToResource

Description

Attaches one or more tags to an existing resource.

Creating a Request

data AddLFTagsToResource

See: newAddLFTagsToResource smart constructor.

Constructors

AddLFTagsToResource'

Fields

catalogId :: Maybe Text

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

resource :: Resource

The resource to which to attach a tag.

lFTags :: NonEmpty LFTagPair

The tags to attach to the resource.

newAddLFTagsToResource

Arguments

:: Resource
$sel:resource:AddLFTagsToResource'

-> NonEmpty LFTagPair
$sel:lFTags:AddLFTagsToResource'

-> AddLFTagsToResource

Create a value of AddLFTagsToResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:AddLFTagsToResource', addLFTagsToResource_catalogId - The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

$sel:resource:AddLFTagsToResource', addLFTagsToResource_resource - The resource to which to attach a tag.

$sel:lFTags:AddLFTagsToResource', addLFTagsToResource_lFTags - The tags to attach to the resource.

Request Lenses

addLFTagsToResource_catalogId :: Lens' AddLFTagsToResource (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

addLFTagsToResource_resource :: Lens' AddLFTagsToResource Resource

The resource to which to attach a tag.

addLFTagsToResource_lFTags :: Lens' AddLFTagsToResource (NonEmpty LFTagPair)

The tags to attach to the resource.

Destructuring the Response

data AddLFTagsToResourceResponse

See: newAddLFTagsToResourceResponse smart constructor.

Constructors

AddLFTagsToResourceResponse'

Fields

failures :: Maybe [LFTagError]

A list of failures to tag the resource.

httpStatus :: Int

The response's http status code.

newAddLFTagsToResourceResponse

Arguments

:: Int
$sel:httpStatus:AddLFTagsToResourceResponse'

-> AddLFTagsToResourceResponse

Create a value of AddLFTagsToResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failures:AddLFTagsToResourceResponse', addLFTagsToResourceResponse_failures - A list of failures to tag the resource.

$sel:httpStatus:AddLFTagsToResourceResponse', addLFTagsToResourceResponse_httpStatus - The response's http status code.

Response Lenses

addLFTagsToResourceResponse_failures :: Lens' AddLFTagsToResourceResponse (Maybe [LFTagError])

A list of failures to tag the resource.

addLFTagsToResourceResponse_httpStatus :: Lens' AddLFTagsToResourceResponse Int

The response's http status code.