Network.AWS.LakeFormation.BatchRevokePermissions

Description

Batch operation to revoke permissions from the principal.

Creating a Request

data BatchRevokePermissions

See: newBatchRevokePermissions smart constructor.

Constructors

BatchRevokePermissions'

Fields

catalogId :: Maybe Text

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

entries :: [BatchPermissionsRequestEntry]

A list of up to 20 entries for resource permissions to be revoked by batch operation to the principal.

newBatchRevokePermissions :: BatchRevokePermissions

Create a value of BatchRevokePermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:BatchRevokePermissions', batchRevokePermissions_catalogId - The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

$sel:entries:BatchRevokePermissions', batchRevokePermissions_entries - A list of up to 20 entries for resource permissions to be revoked by batch operation to the principal.

Request Lenses

batchRevokePermissions_catalogId :: Lens' BatchRevokePermissions (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

batchRevokePermissions_entries :: Lens' BatchRevokePermissions [BatchPermissionsRequestEntry]

A list of up to 20 entries for resource permissions to be revoked by batch operation to the principal.

Destructuring the Response

data BatchRevokePermissionsResponse

See: newBatchRevokePermissionsResponse smart constructor.

Constructors

BatchRevokePermissionsResponse'

Fields

failures :: Maybe [BatchPermissionsFailureEntry]

A list of failures to revoke permissions to the resources.

httpStatus :: Int

The response's http status code.

newBatchRevokePermissionsResponse

Arguments

:: Int
$sel:httpStatus:BatchRevokePermissionsResponse'

-> BatchRevokePermissionsResponse

Create a value of BatchRevokePermissionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failures:BatchRevokePermissionsResponse', batchRevokePermissionsResponse_failures - A list of failures to revoke permissions to the resources.

$sel:httpStatus:BatchRevokePermissionsResponse', batchRevokePermissionsResponse_httpStatus - The response's http status code.

Response Lenses

batchRevokePermissionsResponse_failures :: Lens' BatchRevokePermissionsResponse (Maybe [BatchPermissionsFailureEntry])

A list of failures to revoke permissions to the resources.

batchRevokePermissionsResponse_httpStatus :: Lens' BatchRevokePermissionsResponse Int

The response's http status code.