Network.AWS.LakeFormation.GrantPermissions

Description

Grants permissions to the principal to access metadata in the Data Catalog and data organized in underlying data storage such as Amazon S3.

For information about permissions, see Security and Access Control to Metadata and Data.

Creating a Request

data GrantPermissions

See: newGrantPermissions smart constructor.

Constructors

GrantPermissions'

Fields

catalogId :: Maybe Text

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

permissionsWithGrantOption :: Maybe [Permission]

Indicates a list of the granted permissions that the principal may pass to other users. These permissions may only be a subset of the permissions granted in the Privileges.

principal :: DataLakePrincipal

The principal to be granted the permissions on the resource. Supported principals are IAM users or IAM roles, and they are defined by their principal type and their ARN.

Note that if you define a resource with a particular ARN, then later delete, and recreate a resource with that same ARN, the resource maintains the permissions already granted.

resource :: Resource

The resource to which permissions are to be granted. Resources in AWS Lake Formation are the Data Catalog, databases, and tables.

permissions :: [Permission]

The permissions granted to the principal on the resource. AWS Lake Formation defines privileges to grant and revoke access to metadata in the Data Catalog and data organized in underlying data storage such as Amazon S3. AWS Lake Formation requires that each principal be authorized to perform a specific task on AWS Lake Formation resources.

newGrantPermissions

Arguments

:: DataLakePrincipal
$sel:principal:GrantPermissions'

-> Resource
$sel:resource:GrantPermissions'

-> GrantPermissions

Create a value of GrantPermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:GrantPermissions', grantPermissions_catalogId - The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

$sel:permissionsWithGrantOption:GrantPermissions', grantPermissions_permissionsWithGrantOption - Indicates a list of the granted permissions that the principal may pass to other users. These permissions may only be a subset of the permissions granted in the Privileges.

$sel:principal:GrantPermissions', grantPermissions_principal - The principal to be granted the permissions on the resource. Supported principals are IAM users or IAM roles, and they are defined by their principal type and their ARN.

Note that if you define a resource with a particular ARN, then later delete, and recreate a resource with that same ARN, the resource maintains the permissions already granted.

$sel:resource:GrantPermissions', grantPermissions_resource - The resource to which permissions are to be granted. Resources in AWS Lake Formation are the Data Catalog, databases, and tables.

$sel:permissions:GrantPermissions', grantPermissions_permissions - The permissions granted to the principal on the resource. AWS Lake Formation defines privileges to grant and revoke access to metadata in the Data Catalog and data organized in underlying data storage such as Amazon S3. AWS Lake Formation requires that each principal be authorized to perform a specific task on AWS Lake Formation resources.

Request Lenses

grantPermissions_catalogId :: Lens' GrantPermissions (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

grantPermissions_permissionsWithGrantOption :: Lens' GrantPermissions (Maybe [Permission])

Indicates a list of the granted permissions that the principal may pass to other users. These permissions may only be a subset of the permissions granted in the Privileges.

grantPermissions_principal :: Lens' GrantPermissions DataLakePrincipal

The principal to be granted the permissions on the resource. Supported principals are IAM users or IAM roles, and they are defined by their principal type and their ARN.

Note that if you define a resource with a particular ARN, then later delete, and recreate a resource with that same ARN, the resource maintains the permissions already granted.

grantPermissions_resource :: Lens' GrantPermissions Resource

The resource to which permissions are to be granted. Resources in AWS Lake Formation are the Data Catalog, databases, and tables.

grantPermissions_permissions :: Lens' GrantPermissions [Permission]

The permissions granted to the principal on the resource. AWS Lake Formation defines privileges to grant and revoke access to metadata in the Data Catalog and data organized in underlying data storage such as Amazon S3. AWS Lake Formation requires that each principal be authorized to perform a specific task on AWS Lake Formation resources.

Destructuring the Response

data GrantPermissionsResponse

See: newGrantPermissionsResponse smart constructor.

Constructors

GrantPermissionsResponse'

Fields

httpStatus :: Int

The response's http status code.

newGrantPermissionsResponse

Arguments

:: Int
$sel:httpStatus:GrantPermissionsResponse'

-> GrantPermissionsResponse

Create a value of GrantPermissionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GrantPermissionsResponse', grantPermissionsResponse_httpStatus - The response's http status code.

Response Lenses

grantPermissionsResponse_httpStatus :: Lens' GrantPermissionsResponse Int

The response's http status code.