Network.AWS.LakeFormation.ListPermissions

Description

Returns a list of the principal permissions on the resource, filtered by the permissions of the caller. For example, if you are granted an ALTER permission, you are able to see only the principal permissions for ALTER.

This operation returns only those permissions that have been explicitly granted.

For information about permissions, see Security and Access Control to Metadata and Data.

Creating a Request

data ListPermissions

See: newListPermissions smart constructor.

Constructors

ListPermissions'

Fields

resourceType :: Maybe DataLakeResourceType

Specifies a resource type to filter the permissions returned.

catalogId :: Maybe Text

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

nextToken :: Maybe Text

A continuation token, if this is not the first call to retrieve this list.

principal :: Maybe DataLakePrincipal

Specifies a principal to filter the permissions returned.

resource :: Maybe Resource

A resource where you will get a list of the principal permissions.

This operation does not support getting privileges on a table with columns. Instead, call this operation on the table, and the operation returns the table and the table w columns.

maxResults :: Maybe Natural

The maximum number of results to return.

newListPermissions :: ListPermissions

Create a value of ListPermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:ListPermissions', listPermissions_resourceType - Specifies a resource type to filter the permissions returned.

$sel:catalogId:ListPermissions', listPermissions_catalogId - The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

$sel:nextToken:ListPermissions', listPermissions_nextToken - A continuation token, if this is not the first call to retrieve this list.

$sel:principal:ListPermissions', listPermissions_principal - Specifies a principal to filter the permissions returned.

$sel:resource:ListPermissions', listPermissions_resource - A resource where you will get a list of the principal permissions.

This operation does not support getting privileges on a table with columns. Instead, call this operation on the table, and the operation returns the table and the table w columns.

$sel:maxResults:ListPermissions', listPermissions_maxResults - The maximum number of results to return.

Request Lenses

listPermissions_resourceType :: Lens' ListPermissions (Maybe DataLakeResourceType)

Specifies a resource type to filter the permissions returned.

listPermissions_catalogId :: Lens' ListPermissions (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

listPermissions_nextToken :: Lens' ListPermissions (Maybe Text)

A continuation token, if this is not the first call to retrieve this list.

listPermissions_principal :: Lens' ListPermissions (Maybe DataLakePrincipal)

Specifies a principal to filter the permissions returned.

listPermissions_resource :: Lens' ListPermissions (Maybe Resource)

A resource where you will get a list of the principal permissions.

This operation does not support getting privileges on a table with columns. Instead, call this operation on the table, and the operation returns the table and the table w columns.

listPermissions_maxResults :: Lens' ListPermissions (Maybe Natural)

The maximum number of results to return.

Destructuring the Response

data ListPermissionsResponse

See: newListPermissionsResponse smart constructor.

Constructors

ListPermissionsResponse'

Fields

nextToken :: Maybe Text

A continuation token, if this is not the first call to retrieve this list.

principalResourcePermissions :: Maybe [PrincipalResourcePermissions]

A list of principals and their permissions on the resource for the specified principal and resource types.

httpStatus :: Int

The response's http status code.

newListPermissionsResponse

Arguments

:: Int
$sel:httpStatus:ListPermissionsResponse'

-> ListPermissionsResponse

Create a value of ListPermissionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPermissions', listPermissionsResponse_nextToken - A continuation token, if this is not the first call to retrieve this list.

$sel:principalResourcePermissions:ListPermissionsResponse', listPermissionsResponse_principalResourcePermissions - A list of principals and their permissions on the resource for the specified principal and resource types.

$sel:httpStatus:ListPermissionsResponse', listPermissionsResponse_httpStatus - The response's http status code.

Response Lenses

listPermissionsResponse_nextToken :: Lens' ListPermissionsResponse (Maybe Text)

A continuation token, if this is not the first call to retrieve this list.

listPermissionsResponse_principalResourcePermissions :: Lens' ListPermissionsResponse (Maybe [PrincipalResourcePermissions])

A list of principals and their permissions on the resource for the specified principal and resource types.

listPermissionsResponse_httpStatus :: Lens' ListPermissionsResponse Int

The response's http status code.