Network.AWS.LakeFormation.Types

Description

Service Configuration

defaultService :: Service

API version 2017-03-31 of the Amazon Lake Formation SDK configuration.

Errors

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

Access to a resource was denied.

_OperationTimeoutException :: AsError a => Getting (First ServiceError) a ServiceError

The operation timed out.

_EntityNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

A specified entity does not exist

_ConcurrentModificationException :: AsError a => Getting (First ServiceError) a ServiceError

Two processes are trying to modify a resource simultaneously.

_InternalServiceException :: AsError a => Getting (First ServiceError) a ServiceError

An internal service error occurred.

_InvalidInputException :: AsError a => Getting (First ServiceError) a ServiceError

The input provided was not valid.

_ResourceNumberLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

A resource numerical limit was exceeded.

_GlueEncryptionException :: AsError a => Getting (First ServiceError) a ServiceError

An encryption operation failed.

_AlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

A resource to be created or added already exists.

ComparisonOperator

newtype ComparisonOperator

Constructors

ComparisonOperator'

Fields

fromComparisonOperator :: Text

Bundled Patterns

pattern ComparisonOperator_BEGINS_WITH :: ComparisonOperator

pattern ComparisonOperator_BETWEEN :: ComparisonOperator

pattern ComparisonOperator_CONTAINS :: ComparisonOperator

pattern ComparisonOperator_EQ :: ComparisonOperator

pattern ComparisonOperator_GE :: ComparisonOperator

pattern ComparisonOperator_GT :: ComparisonOperator

pattern ComparisonOperator_IN :: ComparisonOperator

pattern ComparisonOperator_LE :: ComparisonOperator

pattern ComparisonOperator_LT :: ComparisonOperator

pattern ComparisonOperator_NE :: ComparisonOperator

pattern ComparisonOperator_NOT_CONTAINS :: ComparisonOperator

DataLakeResourceType

newtype DataLakeResourceType

Constructors

DataLakeResourceType'

Fields

fromDataLakeResourceType :: Text

Bundled Patterns

pattern DataLakeResourceType_CATALOG :: DataLakeResourceType

pattern DataLakeResourceType_DATABASE :: DataLakeResourceType

pattern DataLakeResourceType_DATA_LOCATION :: DataLakeResourceType

pattern DataLakeResourceType_LF_TAG :: DataLakeResourceType

pattern DataLakeResourceType_LF_TAG_POLICY :: DataLakeResourceType

pattern DataLakeResourceType_LF_TAG_POLICY_DATABASE :: DataLakeResourceType

pattern DataLakeResourceType_LF_TAG_POLICY_TABLE :: DataLakeResourceType

pattern DataLakeResourceType_TABLE :: DataLakeResourceType

FieldNameString

newtype FieldNameString

Constructors

FieldNameString'

Fields

fromFieldNameString :: Text

Bundled Patterns

pattern FieldNameString_LAST_MODIFIED :: FieldNameString

pattern FieldNameString_RESOURCE_ARN :: FieldNameString

pattern FieldNameString_ROLE_ARN :: FieldNameString

Permission

newtype Permission

Constructors

Permission'

Fields

fromPermission :: Text

Bundled Patterns

pattern Permission_ALL :: Permission

pattern Permission_ALTER :: Permission

pattern Permission_ALTER_TAG :: Permission

pattern Permission_ASSOCIATE_TAG :: Permission

pattern Permission_CREATE_DATABASE :: Permission

pattern Permission_CREATE_TABLE :: Permission

pattern Permission_CREATE_TAG :: Permission

pattern Permission_DATA_LOCATION_ACCESS :: Permission

pattern Permission_DELETE :: Permission

pattern Permission_DELETE_TAG :: Permission

pattern Permission_DESCRIBE :: Permission

pattern Permission_DESCRIBE_TAG :: Permission

pattern Permission_DROP :: Permission

pattern Permission_INSERT :: Permission

pattern Permission_SELECT :: Permission

ResourceShareType

newtype ResourceShareType

Constructors

ResourceShareType'

Fields

fromResourceShareType :: Text

Bundled Patterns

pattern ResourceShareType_ALL :: ResourceShareType

pattern ResourceShareType_FOREIGN :: ResourceShareType

ResourceType

newtype ResourceType

Constructors

ResourceType'

Fields

fromResourceType :: Text

Bundled Patterns

pattern ResourceType_DATABASE :: ResourceType

pattern ResourceType_TABLE :: ResourceType

BatchPermissionsFailureEntry

data BatchPermissionsFailureEntry

A list of failures when performing a batch grant or batch revoke operation.

See: newBatchPermissionsFailureEntry smart constructor.

Constructors

BatchPermissionsFailureEntry'

Fields

error :: Maybe ErrorDetail

An error message that applies to the failure of the entry.

requestEntry :: Maybe BatchPermissionsRequestEntry

An identifier for an entry of the batch request.

newBatchPermissionsFailureEntry :: BatchPermissionsFailureEntry

Create a value of BatchPermissionsFailureEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:error:BatchPermissionsFailureEntry', batchPermissionsFailureEntry_error - An error message that applies to the failure of the entry.

$sel:requestEntry:BatchPermissionsFailureEntry', batchPermissionsFailureEntry_requestEntry - An identifier for an entry of the batch request.

batchPermissionsFailureEntry_error :: Lens' BatchPermissionsFailureEntry (Maybe ErrorDetail)

An error message that applies to the failure of the entry.

batchPermissionsFailureEntry_requestEntry :: Lens' BatchPermissionsFailureEntry (Maybe BatchPermissionsRequestEntry)

An identifier for an entry of the batch request.

BatchPermissionsRequestEntry

data BatchPermissionsRequestEntry

A permission to a resource granted by batch operation to the principal.

See: newBatchPermissionsRequestEntry smart constructor.

Constructors

BatchPermissionsRequestEntry'

Fields

permissionsWithGrantOption :: Maybe [Permission]

Indicates if the option to pass permissions is granted.

principal :: Maybe DataLakePrincipal

The principal to be granted a permission.

resource :: Maybe Resource

The resource to which the principal is to be granted a permission.

permissions :: Maybe [Permission]

The permissions to be granted.

id :: Text

A unique identifier for the batch permissions request entry.

newBatchPermissionsRequestEntry

Arguments

:: Text
$sel:id:BatchPermissionsRequestEntry'

-> BatchPermissionsRequestEntry

Create a value of BatchPermissionsRequestEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:permissionsWithGrantOption:BatchPermissionsRequestEntry', batchPermissionsRequestEntry_permissionsWithGrantOption - Indicates if the option to pass permissions is granted.

$sel:principal:BatchPermissionsRequestEntry', batchPermissionsRequestEntry_principal - The principal to be granted a permission.

$sel:resource:BatchPermissionsRequestEntry', batchPermissionsRequestEntry_resource - The resource to which the principal is to be granted a permission.

$sel:permissions:BatchPermissionsRequestEntry', batchPermissionsRequestEntry_permissions - The permissions to be granted.

$sel:id:BatchPermissionsRequestEntry', batchPermissionsRequestEntry_id - A unique identifier for the batch permissions request entry.

batchPermissionsRequestEntry_permissionsWithGrantOption :: Lens' BatchPermissionsRequestEntry (Maybe [Permission])

Indicates if the option to pass permissions is granted.

batchPermissionsRequestEntry_principal :: Lens' BatchPermissionsRequestEntry (Maybe DataLakePrincipal)

The principal to be granted a permission.

batchPermissionsRequestEntry_resource :: Lens' BatchPermissionsRequestEntry (Maybe Resource)

The resource to which the principal is to be granted a permission.

batchPermissionsRequestEntry_permissions :: Lens' BatchPermissionsRequestEntry (Maybe [Permission])

The permissions to be granted.

batchPermissionsRequestEntry_id :: Lens' BatchPermissionsRequestEntry Text

A unique identifier for the batch permissions request entry.

CatalogResource

data CatalogResource

A structure for the catalog object.

See: newCatalogResource smart constructor.

Constructors

CatalogResource'

newCatalogResource :: CatalogResource

Create a value of CatalogResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

ColumnLFTag

data ColumnLFTag

A structure containing the name of a column resource and the tags attached to it.

See: newColumnLFTag smart constructor.

Constructors

ColumnLFTag'

Fields

name :: Maybe Text

The name of a column resource.

lFTags :: Maybe (NonEmpty LFTagPair)

The tags attached to a column resource.

newColumnLFTag :: ColumnLFTag

Create a value of ColumnLFTag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ColumnLFTag', columnLFTag_name - The name of a column resource.

$sel:lFTags:ColumnLFTag', columnLFTag_lFTags - The tags attached to a column resource.

columnLFTag_name :: Lens' ColumnLFTag (Maybe Text)

The name of a column resource.

columnLFTag_lFTags :: Lens' ColumnLFTag (Maybe (NonEmpty LFTagPair))

The tags attached to a column resource.

ColumnWildcard

data ColumnWildcard

A wildcard object, consisting of an optional list of excluded column names or indexes.

See: newColumnWildcard smart constructor.

Constructors

ColumnWildcard'

Fields

excludedColumnNames :: Maybe [Text]

Excludes column names. Any column with this name will be excluded.

newColumnWildcard :: ColumnWildcard

Create a value of ColumnWildcard with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:excludedColumnNames:ColumnWildcard', columnWildcard_excludedColumnNames - Excludes column names. Any column with this name will be excluded.

columnWildcard_excludedColumnNames :: Lens' ColumnWildcard (Maybe [Text])

Excludes column names. Any column with this name will be excluded.

DataLakePrincipal

data DataLakePrincipal

The AWS Lake Formation principal. Supported principals are IAM users or IAM roles.

See: newDataLakePrincipal smart constructor.

Constructors

DataLakePrincipal'

Fields

dataLakePrincipalIdentifier :: Maybe Text

An identifier for the AWS Lake Formation principal.

newDataLakePrincipal :: DataLakePrincipal

Create a value of DataLakePrincipal with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataLakePrincipalIdentifier:DataLakePrincipal', dataLakePrincipal_dataLakePrincipalIdentifier - An identifier for the AWS Lake Formation principal.

dataLakePrincipal_dataLakePrincipalIdentifier :: Lens' DataLakePrincipal (Maybe Text)

An identifier for the AWS Lake Formation principal.

DataLakeSettings

data DataLakeSettings

A structure representing a list of AWS Lake Formation principals designated as data lake administrators and lists of principal permission entries for default create database and default create table permissions.

See: newDataLakeSettings smart constructor.

Constructors

DataLakeSettings'

Fields

dataLakeAdmins :: Maybe [DataLakePrincipal]

A list of AWS Lake Formation principals. Supported principals are IAM users or IAM roles.

trustedResourceOwners :: Maybe [Text]

A list of the resource-owning account IDs that the caller's account can use to share their user access details (user ARNs). The user ARNs can be logged in the resource owner's AWS CloudTrail log.

You may want to specify this property when you are in a high-trust boundary, such as the same team or company.

createDatabaseDefaultPermissions :: Maybe [PrincipalPermissions]

A structure representing a list of up to three principal permissions entries for default create database permissions.

createTableDefaultPermissions :: Maybe [PrincipalPermissions]

A structure representing a list of up to three principal permissions entries for default create table permissions.

newDataLakeSettings :: DataLakeSettings

Create a value of DataLakeSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataLakeAdmins:DataLakeSettings', dataLakeSettings_dataLakeAdmins - A list of AWS Lake Formation principals. Supported principals are IAM users or IAM roles.

$sel:trustedResourceOwners:DataLakeSettings', dataLakeSettings_trustedResourceOwners - A list of the resource-owning account IDs that the caller's account can use to share their user access details (user ARNs). The user ARNs can be logged in the resource owner's AWS CloudTrail log.

You may want to specify this property when you are in a high-trust boundary, such as the same team or company.

$sel:createDatabaseDefaultPermissions:DataLakeSettings', dataLakeSettings_createDatabaseDefaultPermissions - A structure representing a list of up to three principal permissions entries for default create database permissions.

$sel:createTableDefaultPermissions:DataLakeSettings', dataLakeSettings_createTableDefaultPermissions - A structure representing a list of up to three principal permissions entries for default create table permissions.

dataLakeSettings_dataLakeAdmins :: Lens' DataLakeSettings (Maybe [DataLakePrincipal])

A list of AWS Lake Formation principals. Supported principals are IAM users or IAM roles.

dataLakeSettings_trustedResourceOwners :: Lens' DataLakeSettings (Maybe [Text])

A list of the resource-owning account IDs that the caller's account can use to share their user access details (user ARNs). The user ARNs can be logged in the resource owner's AWS CloudTrail log.

You may want to specify this property when you are in a high-trust boundary, such as the same team or company.

dataLakeSettings_createDatabaseDefaultPermissions :: Lens' DataLakeSettings (Maybe [PrincipalPermissions])

A structure representing a list of up to three principal permissions entries for default create database permissions.

dataLakeSettings_createTableDefaultPermissions :: Lens' DataLakeSettings (Maybe [PrincipalPermissions])

A structure representing a list of up to three principal permissions entries for default create table permissions.

DataLocationResource

data DataLocationResource

A structure for a data location object where permissions are granted or revoked.

See: newDataLocationResource smart constructor.

Constructors

DataLocationResource'

Fields

catalogId :: Maybe Text

The identifier for the Data Catalog where the location is registered with AWS Lake Formation. By default, it is the account ID of the caller.

resourceArn :: Text

The Amazon Resource Name (ARN) that uniquely identifies the data location resource.

newDataLocationResource

Arguments

:: Text
$sel:resourceArn:DataLocationResource'

-> DataLocationResource

Create a value of DataLocationResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:DataLocationResource', dataLocationResource_catalogId - The identifier for the Data Catalog where the location is registered with AWS Lake Formation. By default, it is the account ID of the caller.

$sel:resourceArn:DataLocationResource', dataLocationResource_resourceArn - The Amazon Resource Name (ARN) that uniquely identifies the data location resource.

dataLocationResource_catalogId :: Lens' DataLocationResource (Maybe Text)

The identifier for the Data Catalog where the location is registered with AWS Lake Formation. By default, it is the account ID of the caller.

dataLocationResource_resourceArn :: Lens' DataLocationResource Text

The Amazon Resource Name (ARN) that uniquely identifies the data location resource.

DatabaseResource

data DatabaseResource

A structure for the database object.

See: newDatabaseResource smart constructor.

Constructors

DatabaseResource'

Fields

catalogId :: Maybe Text

The identifier for the Data Catalog. By default, it is the account ID of the caller.

name :: Text

The name of the database resource. Unique to the Data Catalog.

newDatabaseResource

Arguments

:: Text
$sel:name:DatabaseResource'

-> DatabaseResource

Create a value of DatabaseResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:DatabaseResource', databaseResource_catalogId - The identifier for the Data Catalog. By default, it is the account ID of the caller.

$sel:name:DatabaseResource', databaseResource_name - The name of the database resource. Unique to the Data Catalog.

databaseResource_catalogId :: Lens' DatabaseResource (Maybe Text)

The identifier for the Data Catalog. By default, it is the account ID of the caller.

databaseResource_name :: Lens' DatabaseResource Text

The name of the database resource. Unique to the Data Catalog.

DetailsMap

data DetailsMap

A structure containing the additional details to be returned in the AdditionalDetails attribute of PrincipalResourcePermissions.

If a catalog resource is shared through AWS Resource Access Manager (AWS RAM), then there will exist a corresponding RAM resource share ARN.

See: newDetailsMap smart constructor.

Constructors

DetailsMap'

Fields

resourceShare :: Maybe [Text]

A resource share ARN for a catalog resource shared through AWS Resource Access Manager (AWS RAM).

newDetailsMap :: DetailsMap

Create a value of DetailsMap with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceShare:DetailsMap', detailsMap_resourceShare - A resource share ARN for a catalog resource shared through AWS Resource Access Manager (AWS RAM).

detailsMap_resourceShare :: Lens' DetailsMap (Maybe [Text])

A resource share ARN for a catalog resource shared through AWS Resource Access Manager (AWS RAM).

ErrorDetail

data ErrorDetail

Contains details about an error.

See: newErrorDetail smart constructor.

Constructors

ErrorDetail'

Fields

errorCode :: Maybe Text

The code associated with this error.

errorMessage :: Maybe Text

A message describing the error.

newErrorDetail :: ErrorDetail

Create a value of ErrorDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorCode:ErrorDetail', errorDetail_errorCode - The code associated with this error.

$sel:errorMessage:ErrorDetail', errorDetail_errorMessage - A message describing the error.

errorDetail_errorCode :: Lens' ErrorDetail (Maybe Text)

The code associated with this error.

errorDetail_errorMessage :: Lens' ErrorDetail (Maybe Text)

A message describing the error.

FilterCondition

data FilterCondition

This structure describes the filtering of columns in a table based on a filter condition.

See: newFilterCondition smart constructor.

Constructors

FilterCondition'

Fields

field :: Maybe FieldNameString

The field to filter in the filter condition.

comparisonOperator :: Maybe ComparisonOperator

The comparison operator used in the filter condition.

stringValueList :: Maybe [Text]

A string with values used in evaluating the filter condition.

newFilterCondition :: FilterCondition

Create a value of FilterCondition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:field:FilterCondition', filterCondition_field - The field to filter in the filter condition.

$sel:comparisonOperator:FilterCondition', filterCondition_comparisonOperator - The comparison operator used in the filter condition.

$sel:stringValueList:FilterCondition', filterCondition_stringValueList - A string with values used in evaluating the filter condition.

filterCondition_field :: Lens' FilterCondition (Maybe FieldNameString)

The field to filter in the filter condition.

filterCondition_comparisonOperator :: Lens' FilterCondition (Maybe ComparisonOperator)

The comparison operator used in the filter condition.

filterCondition_stringValueList :: Lens' FilterCondition (Maybe [Text])

A string with values used in evaluating the filter condition.

LFTag

data LFTag

A structure that allows an admin to grant user permissions on certain conditions. For example, granting a role access to all columns not tagged 'PII' of tables tagged 'Prod'.

See: newLFTag smart constructor.

Constructors

LFTag'

Fields

tagKey :: Text

The key-name for the tag.

tagValues :: NonEmpty Text

A list of possible values an attribute can take.

newLFTag

Arguments

:: Text
$sel:tagKey:LFTag'

-> NonEmpty Text
$sel:tagValues:LFTag'

-> LFTag

Create a value of LFTag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagKey:LFTag', lFTag_tagKey - The key-name for the tag.

$sel:tagValues:LFTag', lFTag_tagValues - A list of possible values an attribute can take.

lFTag_tagKey :: Lens' LFTag Text

The key-name for the tag.

lFTag_tagValues :: Lens' LFTag (NonEmpty Text)

A list of possible values an attribute can take.

LFTagError

data LFTagError

A structure containing an error related to a TagResource or UnTagResource operation.

See: newLFTagError smart constructor.

Constructors

LFTagError'

Fields

lFTag :: Maybe LFTagPair

The key-name of the tag.

error :: Maybe ErrorDetail

An error that occurred with the attachment or detachment of the tag.

newLFTagError :: LFTagError

Create a value of LFTagError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lFTag:LFTagError', lFTagError_lFTag - The key-name of the tag.

$sel:error:LFTagError', lFTagError_error - An error that occurred with the attachment or detachment of the tag.

lFTagError_lFTag :: Lens' LFTagError (Maybe LFTagPair)

The key-name of the tag.

lFTagError_error :: Lens' LFTagError (Maybe ErrorDetail)

An error that occurred with the attachment or detachment of the tag.

LFTagKeyResource

data LFTagKeyResource

A structure containing a tag key and values for a resource.

See: newLFTagKeyResource smart constructor.

Constructors

LFTagKeyResource'

Fields

catalogId :: Maybe Text

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

tagKey :: Text

The key-name for the tag.

tagValues :: NonEmpty Text

A list of possible values an attribute can take.

newLFTagKeyResource

Arguments

:: Text
$sel:tagKey:LFTagKeyResource'

-> NonEmpty Text
$sel:tagValues:LFTagKeyResource'

-> LFTagKeyResource

Create a value of LFTagKeyResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:LFTagKeyResource', lFTagKeyResource_catalogId - The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

$sel:tagKey:LFTagKeyResource', lFTagKeyResource_tagKey - The key-name for the tag.

$sel:tagValues:LFTagKeyResource', lFTagKeyResource_tagValues - A list of possible values an attribute can take.

lFTagKeyResource_catalogId :: Lens' LFTagKeyResource (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

lFTagKeyResource_tagKey :: Lens' LFTagKeyResource Text

The key-name for the tag.

lFTagKeyResource_tagValues :: Lens' LFTagKeyResource (NonEmpty Text)

A list of possible values an attribute can take.

LFTagPair

data LFTagPair

A structure containing a tag key-value pair.

See: newLFTagPair smart constructor.

Constructors

LFTagPair'

Fields

catalogId :: Maybe Text

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

tagKey :: Text

The key-name for the tag.

tagValues :: NonEmpty Text

A list of possible values an attribute can take.

newLFTagPair

Arguments

:: Text
$sel:tagKey:LFTagPair'

-> NonEmpty Text
$sel:tagValues:LFTagPair'

-> LFTagPair

Create a value of LFTagPair with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:LFTagPair', lFTagPair_catalogId - The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

$sel:tagKey:LFTagPair', lFTagPair_tagKey - The key-name for the tag.

$sel:tagValues:LFTagPair', lFTagPair_tagValues - A list of possible values an attribute can take.

lFTagPair_catalogId :: Lens' LFTagPair (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

lFTagPair_tagKey :: Lens' LFTagPair Text

The key-name for the tag.

lFTagPair_tagValues :: Lens' LFTagPair (NonEmpty Text)

A list of possible values an attribute can take.

LFTagPolicyResource

data LFTagPolicyResource

A structure containing a list of tag conditions that apply to a resource's tag policy.

See: newLFTagPolicyResource smart constructor.

Constructors

LFTagPolicyResource'

Fields

catalogId :: Maybe Text

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

resourceType :: ResourceType

The resource type for which the tag policy applies.

expression :: NonEmpty LFTag

A list of tag conditions that apply to the resource's tag policy.

newLFTagPolicyResource

Arguments

:: ResourceType
$sel:resourceType:LFTagPolicyResource'

-> NonEmpty LFTag
$sel:expression:LFTagPolicyResource'

-> LFTagPolicyResource

Create a value of LFTagPolicyResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:LFTagPolicyResource', lFTagPolicyResource_catalogId - The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

$sel:resourceType:LFTagPolicyResource', lFTagPolicyResource_resourceType - The resource type for which the tag policy applies.

$sel:expression:LFTagPolicyResource', lFTagPolicyResource_expression - A list of tag conditions that apply to the resource's tag policy.

lFTagPolicyResource_catalogId :: Lens' LFTagPolicyResource (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

lFTagPolicyResource_resourceType :: Lens' LFTagPolicyResource ResourceType

The resource type for which the tag policy applies.

lFTagPolicyResource_expression :: Lens' LFTagPolicyResource (NonEmpty LFTag)

A list of tag conditions that apply to the resource's tag policy.

PrincipalPermissions

data PrincipalPermissions

Permissions granted to a principal.

See: newPrincipalPermissions smart constructor.

Constructors

PrincipalPermissions'

Fields

principal :: Maybe DataLakePrincipal

The principal who is granted permissions.

permissions :: Maybe [Permission]

The permissions that are granted to the principal.

newPrincipalPermissions :: PrincipalPermissions

Create a value of PrincipalPermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:principal:PrincipalPermissions', principalPermissions_principal - The principal who is granted permissions.

$sel:permissions:PrincipalPermissions', principalPermissions_permissions - The permissions that are granted to the principal.

principalPermissions_principal :: Lens' PrincipalPermissions (Maybe DataLakePrincipal)

The principal who is granted permissions.

principalPermissions_permissions :: Lens' PrincipalPermissions (Maybe [Permission])

The permissions that are granted to the principal.

PrincipalResourcePermissions

data PrincipalResourcePermissions

The permissions granted or revoked on a resource.

See: newPrincipalResourcePermissions smart constructor.

Constructors

PrincipalResourcePermissions'

Fields

additionalDetails :: Maybe DetailsMap

This attribute can be used to return any additional details of PrincipalResourcePermissions. Currently returns only as a RAM resource share ARN.

permissionsWithGrantOption :: Maybe [Permission]

Indicates whether to grant the ability to grant permissions (as a subset of permissions granted).

principal :: Maybe DataLakePrincipal

The Data Lake principal to be granted or revoked permissions.

resource :: Maybe Resource

The resource where permissions are to be granted or revoked.

permissions :: Maybe [Permission]

The permissions to be granted or revoked on the resource.

newPrincipalResourcePermissions :: PrincipalResourcePermissions

Create a value of PrincipalResourcePermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:additionalDetails:PrincipalResourcePermissions', principalResourcePermissions_additionalDetails - This attribute can be used to return any additional details of PrincipalResourcePermissions. Currently returns only as a RAM resource share ARN.

$sel:permissionsWithGrantOption:PrincipalResourcePermissions', principalResourcePermissions_permissionsWithGrantOption - Indicates whether to grant the ability to grant permissions (as a subset of permissions granted).

$sel:principal:PrincipalResourcePermissions', principalResourcePermissions_principal - The Data Lake principal to be granted or revoked permissions.

$sel:resource:PrincipalResourcePermissions', principalResourcePermissions_resource - The resource where permissions are to be granted or revoked.

$sel:permissions:PrincipalResourcePermissions', principalResourcePermissions_permissions - The permissions to be granted or revoked on the resource.

principalResourcePermissions_additionalDetails :: Lens' PrincipalResourcePermissions (Maybe DetailsMap)

This attribute can be used to return any additional details of PrincipalResourcePermissions. Currently returns only as a RAM resource share ARN.

principalResourcePermissions_permissionsWithGrantOption :: Lens' PrincipalResourcePermissions (Maybe [Permission])

Indicates whether to grant the ability to grant permissions (as a subset of permissions granted).

principalResourcePermissions_principal :: Lens' PrincipalResourcePermissions (Maybe DataLakePrincipal)

The Data Lake principal to be granted or revoked permissions.

principalResourcePermissions_resource :: Lens' PrincipalResourcePermissions (Maybe Resource)

The resource where permissions are to be granted or revoked.

principalResourcePermissions_permissions :: Lens' PrincipalResourcePermissions (Maybe [Permission])

The permissions to be granted or revoked on the resource.

Resource

data Resource

A structure for the resource.

See: newResource smart constructor.

Constructors

Resource'

Fields

dataLocation :: Maybe DataLocationResource

The location of an Amazon S3 path where permissions are granted or revoked.

database :: Maybe DatabaseResource

The database for the resource. Unique to the Data Catalog. A database is a set of associated table definitions organized into a logical group. You can Grant and Revoke database permissions to a principal.

lFTag :: Maybe LFTagKeyResource

The tag key and values attached to a resource.

catalog :: Maybe CatalogResource

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

lFTagPolicy :: Maybe LFTagPolicyResource

A list of tag conditions that define a resource's tag policy.

table :: Maybe TableResource

The table for the resource. A table is a metadata definition that represents your data. You can Grant and Revoke table privileges to a principal.

tableWithColumns :: Maybe TableWithColumnsResource

The table with columns for the resource. A principal with permissions to this resource can select metadata from the columns of a table in the Data Catalog and the underlying data in Amazon S3.

newResource :: Resource

Create a value of Resource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataLocation:Resource', resource_dataLocation - The location of an Amazon S3 path where permissions are granted or revoked.

$sel:database:Resource', resource_database - The database for the resource. Unique to the Data Catalog. A database is a set of associated table definitions organized into a logical group. You can Grant and Revoke database permissions to a principal.

$sel:lFTag:Resource', resource_lFTag - The tag key and values attached to a resource.

$sel:catalog:Resource', resource_catalog - The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

$sel:lFTagPolicy:Resource', resource_lFTagPolicy - A list of tag conditions that define a resource's tag policy.

$sel:table:Resource', resource_table - The table for the resource. A table is a metadata definition that represents your data. You can Grant and Revoke table privileges to a principal.

$sel:tableWithColumns:Resource', resource_tableWithColumns - The table with columns for the resource. A principal with permissions to this resource can select metadata from the columns of a table in the Data Catalog and the underlying data in Amazon S3.

resource_dataLocation :: Lens' Resource (Maybe DataLocationResource)

The location of an Amazon S3 path where permissions are granted or revoked.

resource_database :: Lens' Resource (Maybe DatabaseResource)

The database for the resource. Unique to the Data Catalog. A database is a set of associated table definitions organized into a logical group. You can Grant and Revoke database permissions to a principal.

resource_lFTag :: Lens' Resource (Maybe LFTagKeyResource)

The tag key and values attached to a resource.

resource_catalog :: Lens' Resource (Maybe CatalogResource)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

resource_lFTagPolicy :: Lens' Resource (Maybe LFTagPolicyResource)

A list of tag conditions that define a resource's tag policy.

resource_table :: Lens' Resource (Maybe TableResource)

The table for the resource. A table is a metadata definition that represents your data. You can Grant and Revoke table privileges to a principal.

resource_tableWithColumns :: Lens' Resource (Maybe TableWithColumnsResource)

The table with columns for the resource. A principal with permissions to this resource can select metadata from the columns of a table in the Data Catalog and the underlying data in Amazon S3.

ResourceInfo

data ResourceInfo

A structure containing information about an AWS Lake Formation resource.

See: newResourceInfo smart constructor.

Constructors

ResourceInfo'

Fields

resourceArn :: Maybe Text

The Amazon Resource Name (ARN) of the resource.

lastModified :: Maybe POSIX

The date and time the resource was last modified.

roleArn :: Maybe Text

The IAM role that registered a resource.

newResourceInfo :: ResourceInfo

Create a value of ResourceInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ResourceInfo', resourceInfo_resourceArn - The Amazon Resource Name (ARN) of the resource.

$sel:lastModified:ResourceInfo', resourceInfo_lastModified - The date and time the resource was last modified.

$sel:roleArn:ResourceInfo', resourceInfo_roleArn - The IAM role that registered a resource.

resourceInfo_resourceArn :: Lens' ResourceInfo (Maybe Text)

The Amazon Resource Name (ARN) of the resource.

resourceInfo_lastModified :: Lens' ResourceInfo (Maybe UTCTime)

The date and time the resource was last modified.

resourceInfo_roleArn :: Lens' ResourceInfo (Maybe Text)

The IAM role that registered a resource.

TableResource

data TableResource

A structure for the table object. A table is a metadata definition that represents your data. You can Grant and Revoke table privileges to a principal.

See: newTableResource smart constructor.

Constructors

TableResource'

Fields

catalogId :: Maybe Text

The identifier for the Data Catalog. By default, it is the account ID of the caller.

tableWildcard :: Maybe TableWildcard

A wildcard object representing every table under a database.

At least one of TableResource$Name or TableResource$TableWildcard is required.

name :: Maybe Text

The name of the table.

databaseName :: Text

The name of the database for the table. Unique to a Data Catalog. A database is a set of associated table definitions organized into a logical group. You can Grant and Revoke database privileges to a principal.

newTableResource

Arguments

:: Text
$sel:databaseName:TableResource'

-> TableResource

Create a value of TableResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:TableResource', tableResource_catalogId - The identifier for the Data Catalog. By default, it is the account ID of the caller.

$sel:tableWildcard:TableResource', tableResource_tableWildcard - A wildcard object representing every table under a database.

At least one of TableResource$Name or TableResource$TableWildcard is required.

$sel:name:TableResource', tableResource_name - The name of the table.

$sel:databaseName:TableResource', tableResource_databaseName - The name of the database for the table. Unique to a Data Catalog. A database is a set of associated table definitions organized into a logical group. You can Grant and Revoke database privileges to a principal.

tableResource_catalogId :: Lens' TableResource (Maybe Text)

The identifier for the Data Catalog. By default, it is the account ID of the caller.

tableResource_tableWildcard :: Lens' TableResource (Maybe TableWildcard)

A wildcard object representing every table under a database.

At least one of TableResource$Name or TableResource$TableWildcard is required.

tableResource_name :: Lens' TableResource (Maybe Text)

The name of the table.

tableResource_databaseName :: Lens' TableResource Text

The name of the database for the table. Unique to a Data Catalog. A database is a set of associated table definitions organized into a logical group. You can Grant and Revoke database privileges to a principal.

TableWildcard

data TableWildcard

A wildcard object representing every table under a database.

See: newTableWildcard smart constructor.

Constructors

TableWildcard'

newTableWildcard :: TableWildcard

Create a value of TableWildcard with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

TableWithColumnsResource

data TableWithColumnsResource

A structure for a table with columns object. This object is only used when granting a SELECT permission.

This object must take a value for at least one of ColumnsNames, ColumnsIndexes, or ColumnsWildcard.

See: newTableWithColumnsResource smart constructor.

Constructors

TableWithColumnsResource'

Fields

catalogId :: Maybe Text

The identifier for the Data Catalog. By default, it is the account ID of the caller.

columnWildcard :: Maybe ColumnWildcard

A wildcard specified by a ColumnWildcard object. At least one of ColumnNames or ColumnWildcard is required.

columnNames :: Maybe [Text]

The list of column names for the table. At least one of ColumnNames or ColumnWildcard is required.

databaseName :: Text

The name of the database for the table with columns resource. Unique to the Data Catalog. A database is a set of associated table definitions organized into a logical group. You can Grant and Revoke database privileges to a principal.

name :: Text

The name of the table resource. A table is a metadata definition that represents your data. You can Grant and Revoke table privileges to a principal.

newTableWithColumnsResource

Arguments

:: Text
$sel:databaseName:TableWithColumnsResource'

-> Text
$sel:name:TableWithColumnsResource'

-> TableWithColumnsResource

Create a value of TableWithColumnsResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:TableWithColumnsResource', tableWithColumnsResource_catalogId - The identifier for the Data Catalog. By default, it is the account ID of the caller.

$sel:columnWildcard:TableWithColumnsResource', tableWithColumnsResource_columnWildcard - A wildcard specified by a ColumnWildcard object. At least one of ColumnNames or ColumnWildcard is required.

$sel:columnNames:TableWithColumnsResource', tableWithColumnsResource_columnNames - The list of column names for the table. At least one of ColumnNames or ColumnWildcard is required.

$sel:databaseName:TableWithColumnsResource', tableWithColumnsResource_databaseName - The name of the database for the table with columns resource. Unique to the Data Catalog. A database is a set of associated table definitions organized into a logical group. You can Grant and Revoke database privileges to a principal.

$sel:name:TableWithColumnsResource', tableWithColumnsResource_name - The name of the table resource. A table is a metadata definition that represents your data. You can Grant and Revoke table privileges to a principal.

tableWithColumnsResource_catalogId :: Lens' TableWithColumnsResource (Maybe Text)

The identifier for the Data Catalog. By default, it is the account ID of the caller.

tableWithColumnsResource_columnWildcard :: Lens' TableWithColumnsResource (Maybe ColumnWildcard)

A wildcard specified by a ColumnWildcard object. At least one of ColumnNames or ColumnWildcard is required.

tableWithColumnsResource_columnNames :: Lens' TableWithColumnsResource (Maybe [Text])

The list of column names for the table. At least one of ColumnNames or ColumnWildcard is required.

tableWithColumnsResource_databaseName :: Lens' TableWithColumnsResource Text

The name of the database for the table with columns resource. Unique to the Data Catalog. A database is a set of associated table definitions organized into a logical group. You can Grant and Revoke database privileges to a principal.

tableWithColumnsResource_name :: Lens' TableWithColumnsResource Text

The name of the table resource. A table is a metadata definition that represents your data. You can Grant and Revoke table privileges to a principal.

TaggedDatabase

data TaggedDatabase

A structure describing a database resource with tags.

See: newTaggedDatabase smart constructor.

Constructors

TaggedDatabase'

Fields

database :: Maybe DatabaseResource

A database that has tags attached to it.

lFTags :: Maybe (NonEmpty LFTagPair)

A list of tags attached to the database.

newTaggedDatabase :: TaggedDatabase

Create a value of TaggedDatabase with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:database:TaggedDatabase', taggedDatabase_database - A database that has tags attached to it.

$sel:lFTags:TaggedDatabase', taggedDatabase_lFTags - A list of tags attached to the database.

taggedDatabase_database :: Lens' TaggedDatabase (Maybe DatabaseResource)

A database that has tags attached to it.

taggedDatabase_lFTags :: Lens' TaggedDatabase (Maybe (NonEmpty LFTagPair))

A list of tags attached to the database.

TaggedTable

data TaggedTable

A structure describing a table resource with tags.

See: newTaggedTable smart constructor.

Constructors

TaggedTable'

Fields

lFTagsOnTable :: Maybe (NonEmpty LFTagPair)

A list of tags attached to the table.

lFTagOnDatabase :: Maybe (NonEmpty LFTagPair)

A list of tags attached to the database where the table resides.

lFTagsOnColumns :: Maybe [ColumnLFTag]

A list of tags attached to columns in the table.

table :: Maybe TableResource

A table that has tags attached to it.

newTaggedTable :: TaggedTable

Create a value of TaggedTable with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lFTagsOnTable:TaggedTable', taggedTable_lFTagsOnTable - A list of tags attached to the table.

$sel:lFTagOnDatabase:TaggedTable', taggedTable_lFTagOnDatabase - A list of tags attached to the database where the table resides.

$sel:lFTagsOnColumns:TaggedTable', taggedTable_lFTagsOnColumns - A list of tags attached to columns in the table.

$sel:table:TaggedTable', taggedTable_table - A table that has tags attached to it.

taggedTable_lFTagsOnTable :: Lens' TaggedTable (Maybe (NonEmpty LFTagPair))

A list of tags attached to the table.

taggedTable_lFTagOnDatabase :: Lens' TaggedTable (Maybe (NonEmpty LFTagPair))

A list of tags attached to the database where the table resides.

taggedTable_lFTagsOnColumns :: Lens' TaggedTable (Maybe [ColumnLFTag])

A list of tags attached to columns in the table.

taggedTable_table :: Lens' TaggedTable (Maybe TableResource)

A table that has tags attached to it.