Network.AWS.LakeFormation.Types.LFTagPolicyResource

Description

Documentation

data LFTagPolicyResource

A structure containing a list of tag conditions that apply to a resource's tag policy.

See: newLFTagPolicyResource smart constructor.

Constructors

LFTagPolicyResource'

Fields

catalogId :: Maybe Text

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

resourceType :: ResourceType

The resource type for which the tag policy applies.

expression :: NonEmpty LFTag

A list of tag conditions that apply to the resource's tag policy.

newLFTagPolicyResource

Arguments

:: ResourceType
$sel:resourceType:LFTagPolicyResource'

-> NonEmpty LFTag
$sel:expression:LFTagPolicyResource'

-> LFTagPolicyResource

Create a value of LFTagPolicyResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:LFTagPolicyResource', lFTagPolicyResource_catalogId - The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

$sel:resourceType:LFTagPolicyResource', lFTagPolicyResource_resourceType - The resource type for which the tag policy applies.

$sel:expression:LFTagPolicyResource', lFTagPolicyResource_expression - A list of tag conditions that apply to the resource's tag policy.

lFTagPolicyResource_catalogId :: Lens' LFTagPolicyResource (Maybe Text)

The identifier for the Data Catalog. By default, the account ID. The Data Catalog is the persistent metadata store. It contains database definitions, table definitions, and other control information to manage your AWS Lake Formation environment.

lFTagPolicyResource_resourceType :: Lens' LFTagPolicyResource ResourceType

The resource type for which the tag policy applies.

lFTagPolicyResource_expression :: Lens' LFTagPolicyResource (NonEmpty LFTag)

A list of tag conditions that apply to the resource's tag policy.