Network.AWS.LakeFormation.Types.ResourceInfo

Description

Documentation

data ResourceInfo

A structure containing information about an AWS Lake Formation resource.

See: newResourceInfo smart constructor.

Constructors

ResourceInfo'

Fields

resourceArn :: Maybe Text

The Amazon Resource Name (ARN) of the resource.

lastModified :: Maybe POSIX

The date and time the resource was last modified.

roleArn :: Maybe Text

The IAM role that registered a resource.

newResourceInfo :: ResourceInfo

Create a value of ResourceInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:ResourceInfo', resourceInfo_resourceArn - The Amazon Resource Name (ARN) of the resource.

$sel:lastModified:ResourceInfo', resourceInfo_lastModified - The date and time the resource was last modified.

$sel:roleArn:ResourceInfo', resourceInfo_roleArn - The IAM role that registered a resource.

resourceInfo_resourceArn :: Lens' ResourceInfo (Maybe Text)

The Amazon Resource Name (ARN) of the resource.

resourceInfo_lastModified :: Lens' ResourceInfo (Maybe UTCTime)

The date and time the resource was last modified.

resourceInfo_roleArn :: Lens' ResourceInfo (Maybe Text)

The IAM role that registered a resource.