Network.AWS.LakeFormation.Types.TableWithColumnsResource

Description

Documentation

data TableWithColumnsResource

A structure for a table with columns object. This object is only used when granting a SELECT permission.

This object must take a value for at least one of ColumnsNames, ColumnsIndexes, or ColumnsWildcard.

See: newTableWithColumnsResource smart constructor.

Constructors

TableWithColumnsResource'

Fields

catalogId :: Maybe Text

The identifier for the Data Catalog. By default, it is the account ID of the caller.

columnWildcard :: Maybe ColumnWildcard

A wildcard specified by a ColumnWildcard object. At least one of ColumnNames or ColumnWildcard is required.

columnNames :: Maybe [Text]

The list of column names for the table. At least one of ColumnNames or ColumnWildcard is required.

databaseName :: Text

The name of the database for the table with columns resource. Unique to the Data Catalog. A database is a set of associated table definitions organized into a logical group. You can Grant and Revoke database privileges to a principal.

name :: Text

The name of the table resource. A table is a metadata definition that represents your data. You can Grant and Revoke table privileges to a principal.

newTableWithColumnsResource

Arguments

:: Text
$sel:databaseName:TableWithColumnsResource'

-> Text
$sel:name:TableWithColumnsResource'

-> TableWithColumnsResource

Create a value of TableWithColumnsResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:catalogId:TableWithColumnsResource', tableWithColumnsResource_catalogId - The identifier for the Data Catalog. By default, it is the account ID of the caller.

$sel:columnWildcard:TableWithColumnsResource', tableWithColumnsResource_columnWildcard - A wildcard specified by a ColumnWildcard object. At least one of ColumnNames or ColumnWildcard is required.

$sel:columnNames:TableWithColumnsResource', tableWithColumnsResource_columnNames - The list of column names for the table. At least one of ColumnNames or ColumnWildcard is required.

$sel:databaseName:TableWithColumnsResource', tableWithColumnsResource_databaseName - The name of the database for the table with columns resource. Unique to the Data Catalog. A database is a set of associated table definitions organized into a logical group. You can Grant and Revoke database privileges to a principal.

$sel:name:TableWithColumnsResource', tableWithColumnsResource_name - The name of the table resource. A table is a metadata definition that represents your data. You can Grant and Revoke table privileges to a principal.

tableWithColumnsResource_catalogId :: Lens' TableWithColumnsResource (Maybe Text)

The identifier for the Data Catalog. By default, it is the account ID of the caller.

tableWithColumnsResource_columnWildcard :: Lens' TableWithColumnsResource (Maybe ColumnWildcard)

A wildcard specified by a ColumnWildcard object. At least one of ColumnNames or ColumnWildcard is required.

tableWithColumnsResource_columnNames :: Lens' TableWithColumnsResource (Maybe [Text])

The list of column names for the table. At least one of ColumnNames or ColumnWildcard is required.

tableWithColumnsResource_databaseName :: Lens' TableWithColumnsResource Text

The name of the database for the table with columns resource. Unique to the Data Catalog. A database is a set of associated table definitions organized into a logical group. You can Grant and Revoke database privileges to a principal.

tableWithColumnsResource_name :: Lens' TableWithColumnsResource Text

The name of the table resource. A table is a metadata definition that represents your data. You can Grant and Revoke table privileges to a principal.