Network.AWS.LakeFormation.Types.TaggedTable

Description

Documentation

data TaggedTable

A structure describing a table resource with tags.

See: newTaggedTable smart constructor.

Constructors

TaggedTable'

Fields

lFTagsOnTable :: Maybe (NonEmpty LFTagPair)

A list of tags attached to the table.

lFTagOnDatabase :: Maybe (NonEmpty LFTagPair)

A list of tags attached to the database where the table resides.

lFTagsOnColumns :: Maybe [ColumnLFTag]

A list of tags attached to columns in the table.

table :: Maybe TableResource

A table that has tags attached to it.

newTaggedTable :: TaggedTable

Create a value of TaggedTable with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lFTagsOnTable:TaggedTable', taggedTable_lFTagsOnTable - A list of tags attached to the table.

$sel:lFTagOnDatabase:TaggedTable', taggedTable_lFTagOnDatabase - A list of tags attached to the database where the table resides.

$sel:lFTagsOnColumns:TaggedTable', taggedTable_lFTagsOnColumns - A list of tags attached to columns in the table.

$sel:table:TaggedTable', taggedTable_table - A table that has tags attached to it.

taggedTable_lFTagsOnTable :: Lens' TaggedTable (Maybe (NonEmpty LFTagPair))

A list of tags attached to the table.

taggedTable_lFTagOnDatabase :: Lens' TaggedTable (Maybe (NonEmpty LFTagPair))

A list of tags attached to the database where the table resides.

taggedTable_lFTagsOnColumns :: Lens' TaggedTable (Maybe [ColumnLFTag])

A list of tags attached to columns in the table.

taggedTable_table :: Lens' TaggedTable (Maybe TableResource)

A table that has tags attached to it.