Network.AWS.Lambda.CreateAlias

Description

Creates an alias for a Lambda function version. Use aliases to provide clients with a function identifier that you can update to invoke a different version.

You can also map an alias to split invocation requests between two versions. Use the RoutingConfig parameter to specify a second version and the percentage of invocation requests that it receives.

Creating a Request

data CreateAlias

See: newCreateAlias smart constructor.

Constructors

CreateAlias'

Fields

routingConfig :: Maybe AliasRoutingConfiguration

The routing configuration of the alias.

description :: Maybe Text

A description of the alias.

functionName :: Text

The name of the Lambda function.

Name formats

Function name - MyFunction.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction.

Partial ARN - 123456789012:function:MyFunction.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

name :: Text

The name of the alias.

functionVersion :: Text

The function version that the alias invokes.

newCreateAlias

Arguments

:: Text
$sel:functionName:CreateAlias'

-> Text
$sel:name:CreateAlias'

-> Text
$sel:functionVersion:CreateAlias'

-> CreateAlias

Create a value of CreateAlias with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:routingConfig:CreateAlias', createAlias_routingConfig - The routing configuration of the alias.

$sel:description:CreateAlias', createAlias_description - A description of the alias.

$sel:functionName:CreateAlias', createAlias_functionName - The name of the Lambda function.

Name formats

Function name - MyFunction.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction.

Partial ARN - 123456789012:function:MyFunction.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

$sel:name:CreateAlias', createAlias_name - The name of the alias.

$sel:functionVersion:CreateAlias', createAlias_functionVersion - The function version that the alias invokes.

Request Lenses

createAlias_routingConfig :: Lens' CreateAlias (Maybe AliasRoutingConfiguration)

The routing configuration of the alias.

createAlias_description :: Lens' CreateAlias (Maybe Text)

A description of the alias.

createAlias_functionName :: Lens' CreateAlias Text

The name of the Lambda function.

Name formats

Function name - MyFunction.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction.

Partial ARN - 123456789012:function:MyFunction.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

createAlias_name :: Lens' CreateAlias Text

The name of the alias.

createAlias_functionVersion :: Lens' CreateAlias Text

The function version that the alias invokes.

Destructuring the Response

data AliasConfiguration

Provides configuration information about a Lambda function alias.

See: newAliasConfiguration smart constructor.

Constructors

AliasConfiguration'

Fields

routingConfig :: Maybe AliasRoutingConfiguration

The routing configuration of the alias.

name :: Maybe Text

The name of the alias.

functionVersion :: Maybe Text

The function version that the alias invokes.

aliasArn :: Maybe Text

The Amazon Resource Name (ARN) of the alias.

description :: Maybe Text

A description of the alias.

revisionId :: Maybe Text

A unique identifier that changes when you update the alias.

newAliasConfiguration :: AliasConfiguration

Create a value of AliasConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:routingConfig:AliasConfiguration', aliasConfiguration_routingConfig - The routing configuration of the alias.

$sel:name:AliasConfiguration', aliasConfiguration_name - The name of the alias.

$sel:functionVersion:AliasConfiguration', aliasConfiguration_functionVersion - The function version that the alias invokes.

$sel:aliasArn:AliasConfiguration', aliasConfiguration_aliasArn - The Amazon Resource Name (ARN) of the alias.

$sel:description:AliasConfiguration', aliasConfiguration_description - A description of the alias.

$sel:revisionId:AliasConfiguration', aliasConfiguration_revisionId - A unique identifier that changes when you update the alias.

Response Lenses

aliasConfiguration_routingConfig :: Lens' AliasConfiguration (Maybe AliasRoutingConfiguration)

The routing configuration of the alias.

aliasConfiguration_name :: Lens' AliasConfiguration (Maybe Text)

The name of the alias.

aliasConfiguration_functionVersion :: Lens' AliasConfiguration (Maybe Text)

The function version that the alias invokes.

aliasConfiguration_aliasArn :: Lens' AliasConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the alias.

aliasConfiguration_description :: Lens' AliasConfiguration (Maybe Text)

A description of the alias.

aliasConfiguration_revisionId :: Lens' AliasConfiguration (Maybe Text)

A unique identifier that changes when you update the alias.