Network.AWS.Lambda.CreateCodeSigningConfig

Description

Creates a code signing configuration. A code signing configuration defines a list of allowed signing profiles and defines the code-signing validation policy (action to be taken if deployment validation checks fail).

Creating a Request

data CreateCodeSigningConfig

See: newCreateCodeSigningConfig smart constructor.

Constructors

CreateCodeSigningConfig'

Fields

description :: Maybe Text

Descriptive name for this code signing configuration.

codeSigningPolicies :: Maybe CodeSigningPolicies

The code signing policies define the actions to take if the validation checks fail.

allowedPublishers :: AllowedPublishers

Signing profiles for this code signing configuration.

newCreateCodeSigningConfig

Arguments

:: AllowedPublishers
$sel:allowedPublishers:CreateCodeSigningConfig'

-> CreateCodeSigningConfig

Create a value of CreateCodeSigningConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateCodeSigningConfig', createCodeSigningConfig_description - Descriptive name for this code signing configuration.

$sel:codeSigningPolicies:CreateCodeSigningConfig', createCodeSigningConfig_codeSigningPolicies - The code signing policies define the actions to take if the validation checks fail.

$sel:allowedPublishers:CreateCodeSigningConfig', createCodeSigningConfig_allowedPublishers - Signing profiles for this code signing configuration.

Request Lenses

createCodeSigningConfig_description :: Lens' CreateCodeSigningConfig (Maybe Text)

Descriptive name for this code signing configuration.

createCodeSigningConfig_codeSigningPolicies :: Lens' CreateCodeSigningConfig (Maybe CodeSigningPolicies)

The code signing policies define the actions to take if the validation checks fail.

createCodeSigningConfig_allowedPublishers :: Lens' CreateCodeSigningConfig AllowedPublishers

Signing profiles for this code signing configuration.

Destructuring the Response

data CreateCodeSigningConfigResponse

See: newCreateCodeSigningConfigResponse smart constructor.

Constructors

CreateCodeSigningConfigResponse'

Fields

httpStatus :: Int

The response's http status code.

codeSigningConfig :: CodeSigningConfig

The code signing configuration.

newCreateCodeSigningConfigResponse

Arguments

:: Int
$sel:httpStatus:CreateCodeSigningConfigResponse'

-> CodeSigningConfig
$sel:codeSigningConfig:CreateCodeSigningConfigResponse'

-> CreateCodeSigningConfigResponse

Create a value of CreateCodeSigningConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateCodeSigningConfigResponse', createCodeSigningConfigResponse_httpStatus - The response's http status code.

$sel:codeSigningConfig:CreateCodeSigningConfigResponse', createCodeSigningConfigResponse_codeSigningConfig - The code signing configuration.

Response Lenses

createCodeSigningConfigResponse_httpStatus :: Lens' CreateCodeSigningConfigResponse Int

The response's http status code.

createCodeSigningConfigResponse_codeSigningConfig :: Lens' CreateCodeSigningConfigResponse CodeSigningConfig

The code signing configuration.