Network.AWS.Lambda.GetAccountSettings

Description

Retrieves details about your account's limits and usage in an Amazon Web Services Region.

Creating a Request

data GetAccountSettings

See: newGetAccountSettings smart constructor.

Constructors

GetAccountSettings'

newGetAccountSettings :: GetAccountSettings

Create a value of GetAccountSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data GetAccountSettingsResponse

See: newGetAccountSettingsResponse smart constructor.

Constructors

GetAccountSettingsResponse'

Fields

accountLimit :: Maybe AccountLimit

Limits that are related to concurrency and code storage.

accountUsage :: Maybe AccountUsage

The number of functions and amount of storage in use.

httpStatus :: Int

The response's http status code.

newGetAccountSettingsResponse

Arguments

:: Int
$sel:httpStatus:GetAccountSettingsResponse'

-> GetAccountSettingsResponse

Create a value of GetAccountSettingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountLimit:GetAccountSettingsResponse', getAccountSettingsResponse_accountLimit - Limits that are related to concurrency and code storage.

$sel:accountUsage:GetAccountSettingsResponse', getAccountSettingsResponse_accountUsage - The number of functions and amount of storage in use.

$sel:httpStatus:GetAccountSettingsResponse', getAccountSettingsResponse_httpStatus - The response's http status code.

Response Lenses

getAccountSettingsResponse_accountLimit :: Lens' GetAccountSettingsResponse (Maybe AccountLimit)

Limits that are related to concurrency and code storage.

getAccountSettingsResponse_accountUsage :: Lens' GetAccountSettingsResponse (Maybe AccountUsage)

The number of functions and amount of storage in use.

getAccountSettingsResponse_httpStatus :: Lens' GetAccountSettingsResponse Int

The response's http status code.