Network.AWS.Lambda.GetFunction

Description

Returns information about the function or function version, with a link to download the deployment package that's valid for 10 minutes. If you specify a function version, only details that are specific to that version are returned.

Creating a Request

data GetFunction

See: newGetFunction smart constructor.

Constructors

GetFunction'

Fields

qualifier :: Maybe Text

Specify a version or alias to get details about a published version of the function.

functionName :: Text

The name of the Lambda function, version, or alias.

Name formats

Function name - my-function (name-only), my-function:v1 (with alias).

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

You can append a version number or alias to any of the formats. The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

newGetFunction

Arguments

:: Text
$sel:functionName:GetFunction'

-> GetFunction

Create a value of GetFunction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:qualifier:GetFunction', getFunction_qualifier - Specify a version or alias to get details about a published version of the function.

$sel:functionName:GetFunction', getFunction_functionName - The name of the Lambda function, version, or alias.

Name formats

Function name - my-function (name-only), my-function:v1 (with alias).

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

You can append a version number or alias to any of the formats. The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

Request Lenses

getFunction_qualifier :: Lens' GetFunction (Maybe Text)

Specify a version or alias to get details about a published version of the function.

getFunction_functionName :: Lens' GetFunction Text

The name of the Lambda function, version, or alias.

Name formats

Function name - my-function (name-only), my-function:v1 (with alias).

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

You can append a version number or alias to any of the formats. The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

Destructuring the Response

data GetFunctionResponse

See: newGetFunctionResponse smart constructor.

Constructors

GetFunctionResponse'

Fields

configuration :: Maybe FunctionConfiguration

The configuration of the function or version.

code :: Maybe FunctionCodeLocation

The deployment package of the function or version.

tags :: Maybe (HashMap Text Text)

The function's tags.

concurrency :: Maybe Concurrency

The function's reserved concurrency.

httpStatus :: Int

The response's http status code.

newGetFunctionResponse

Arguments

:: Int
$sel:httpStatus:GetFunctionResponse'

-> GetFunctionResponse

Create a value of GetFunctionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configuration:GetFunctionResponse', getFunctionResponse_configuration - The configuration of the function or version.

$sel:code:GetFunctionResponse', getFunctionResponse_code - The deployment package of the function or version.

$sel:tags:GetFunctionResponse', getFunctionResponse_tags - The function's tags.

$sel:concurrency:GetFunctionResponse', getFunctionResponse_concurrency - The function's reserved concurrency.

$sel:httpStatus:GetFunctionResponse', getFunctionResponse_httpStatus - The response's http status code.

Response Lenses

getFunctionResponse_configuration :: Lens' GetFunctionResponse (Maybe FunctionConfiguration)

The configuration of the function or version.

getFunctionResponse_code :: Lens' GetFunctionResponse (Maybe FunctionCodeLocation)

The deployment package of the function or version.

getFunctionResponse_tags :: Lens' GetFunctionResponse (Maybe (HashMap Text Text))

The function's tags.

getFunctionResponse_concurrency :: Lens' GetFunctionResponse (Maybe Concurrency)

The function's reserved concurrency.

getFunctionResponse_httpStatus :: Lens' GetFunctionResponse Int

The response's http status code.