Network.AWS.Lambda.GetFunctionEventInvokeConfig

Description

Retrieves the configuration for asynchronous invocation for a function, version, or alias.

To configure options for asynchronous invocation, use PutFunctionEventInvokeConfig.

Creating a Request

data GetFunctionEventInvokeConfig

See: newGetFunctionEventInvokeConfig smart constructor.

Constructors

GetFunctionEventInvokeConfig'

Fields

qualifier :: Maybe Text

A version number or alias name.

functionName :: Text

The name of the Lambda function, version, or alias.

Name formats

Function name - my-function (name-only), my-function:v1 (with alias).

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

You can append a version number or alias to any of the formats. The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

newGetFunctionEventInvokeConfig

Arguments

:: Text
$sel:functionName:GetFunctionEventInvokeConfig'

-> GetFunctionEventInvokeConfig

Create a value of GetFunctionEventInvokeConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:qualifier:GetFunctionEventInvokeConfig', getFunctionEventInvokeConfig_qualifier - A version number or alias name.

$sel:functionName:GetFunctionEventInvokeConfig', getFunctionEventInvokeConfig_functionName - The name of the Lambda function, version, or alias.

Name formats

Function name - my-function (name-only), my-function:v1 (with alias).

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

You can append a version number or alias to any of the formats. The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

Request Lenses

getFunctionEventInvokeConfig_qualifier :: Lens' GetFunctionEventInvokeConfig (Maybe Text)

A version number or alias name.

getFunctionEventInvokeConfig_functionName :: Lens' GetFunctionEventInvokeConfig Text

The name of the Lambda function, version, or alias.

Name formats

Function name - my-function (name-only), my-function:v1 (with alias).

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

You can append a version number or alias to any of the formats. The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

Destructuring the Response

data FunctionEventInvokeConfig

See: newFunctionEventInvokeConfig smart constructor.

Constructors

FunctionEventInvokeConfig'

Fields

functionArn :: Maybe Text

The Amazon Resource Name (ARN) of the function.

maximumEventAgeInSeconds :: Maybe Natural

The maximum age of a request that Lambda sends to a function for processing.

maximumRetryAttempts :: Maybe Natural

The maximum number of times to retry when the function returns an error.

lastModified :: Maybe POSIX

The date and time that the configuration was last updated.

destinationConfig :: Maybe DestinationConfig

A destination for events after they have been sent to a function for processing.

Destinations

Function - The Amazon Resource Name (ARN) of a Lambda function.

Queue - The ARN of an SQS queue.

Topic - The ARN of an SNS topic.

Event Bus - The ARN of an Amazon EventBridge event bus.

newFunctionEventInvokeConfig :: FunctionEventInvokeConfig

Create a value of FunctionEventInvokeConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:functionArn:FunctionEventInvokeConfig', functionEventInvokeConfig_functionArn - The Amazon Resource Name (ARN) of the function.

$sel:maximumEventAgeInSeconds:FunctionEventInvokeConfig', functionEventInvokeConfig_maximumEventAgeInSeconds - The maximum age of a request that Lambda sends to a function for processing.

$sel:maximumRetryAttempts:FunctionEventInvokeConfig', functionEventInvokeConfig_maximumRetryAttempts - The maximum number of times to retry when the function returns an error.

$sel:lastModified:FunctionEventInvokeConfig', functionEventInvokeConfig_lastModified - The date and time that the configuration was last updated.

$sel:destinationConfig:FunctionEventInvokeConfig', functionEventInvokeConfig_destinationConfig - A destination for events after they have been sent to a function for processing.

Destinations

Function - The Amazon Resource Name (ARN) of a Lambda function.

Queue - The ARN of an SQS queue.

Topic - The ARN of an SNS topic.

Event Bus - The ARN of an Amazon EventBridge event bus.

Response Lenses

functionEventInvokeConfig_functionArn :: Lens' FunctionEventInvokeConfig (Maybe Text)

The Amazon Resource Name (ARN) of the function.

functionEventInvokeConfig_maximumEventAgeInSeconds :: Lens' FunctionEventInvokeConfig (Maybe Natural)

The maximum age of a request that Lambda sends to a function for processing.

functionEventInvokeConfig_maximumRetryAttempts :: Lens' FunctionEventInvokeConfig (Maybe Natural)

The maximum number of times to retry when the function returns an error.

functionEventInvokeConfig_lastModified :: Lens' FunctionEventInvokeConfig (Maybe UTCTime)

The date and time that the configuration was last updated.

functionEventInvokeConfig_destinationConfig :: Lens' FunctionEventInvokeConfig (Maybe DestinationConfig)

A destination for events after they have been sent to a function for processing.

Destinations

Function - The Amazon Resource Name (ARN) of a Lambda function.

Queue - The ARN of an SQS queue.

Topic - The ARN of an SNS topic.

Event Bus - The ARN of an Amazon EventBridge event bus.