Network.AWS.Lambda.GetLayerVersionPolicy

Description

Returns the permission policy for a version of an Lambda layer. For more information, see AddLayerVersionPermission.

Creating a Request

data GetLayerVersionPolicy

See: newGetLayerVersionPolicy smart constructor.

Constructors

GetLayerVersionPolicy'

Fields

layerName :: Text

The name or Amazon Resource Name (ARN) of the layer.

versionNumber :: Integer

The version number.

newGetLayerVersionPolicy

Arguments

:: Text
$sel:layerName:GetLayerVersionPolicy'

-> Integer
$sel:versionNumber:GetLayerVersionPolicy'

-> GetLayerVersionPolicy

Create a value of GetLayerVersionPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:layerName:GetLayerVersionPolicy', getLayerVersionPolicy_layerName - The name or Amazon Resource Name (ARN) of the layer.

$sel:versionNumber:GetLayerVersionPolicy', getLayerVersionPolicy_versionNumber - The version number.

Request Lenses

getLayerVersionPolicy_layerName :: Lens' GetLayerVersionPolicy Text

The name or Amazon Resource Name (ARN) of the layer.

getLayerVersionPolicy_versionNumber :: Lens' GetLayerVersionPolicy Integer

The version number.

Destructuring the Response

data GetLayerVersionPolicyResponse

See: newGetLayerVersionPolicyResponse smart constructor.

Constructors

GetLayerVersionPolicyResponse'

Fields

revisionId :: Maybe Text

A unique identifier for the current revision of the policy.

policy :: Maybe Text

The policy document.

httpStatus :: Int

The response's http status code.

newGetLayerVersionPolicyResponse

Arguments

:: Int
$sel:httpStatus:GetLayerVersionPolicyResponse'

-> GetLayerVersionPolicyResponse

Create a value of GetLayerVersionPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:revisionId:GetLayerVersionPolicyResponse', getLayerVersionPolicyResponse_revisionId - A unique identifier for the current revision of the policy.

$sel:policy:GetLayerVersionPolicyResponse', getLayerVersionPolicyResponse_policy - The policy document.

$sel:httpStatus:GetLayerVersionPolicyResponse', getLayerVersionPolicyResponse_httpStatus - The response's http status code.

Response Lenses

getLayerVersionPolicyResponse_revisionId :: Lens' GetLayerVersionPolicyResponse (Maybe Text)

A unique identifier for the current revision of the policy.

getLayerVersionPolicyResponse_policy :: Lens' GetLayerVersionPolicyResponse (Maybe Text)

The policy document.

getLayerVersionPolicyResponse_httpStatus :: Lens' GetLayerVersionPolicyResponse Int

The response's http status code.