Network.AWS.Lambda.GetPolicy

Description

Returns the resource-based IAM policy for a function, version, or alias.

Creating a Request

data GetPolicy

See: newGetPolicy smart constructor.

Constructors

GetPolicy'

Fields

qualifier :: Maybe Text

Specify a version or alias to get the policy for that resource.

functionName :: Text

The name of the Lambda function, version, or alias.

Name formats

Function name - my-function (name-only), my-function:v1 (with alias).

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

You can append a version number or alias to any of the formats. The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

newGetPolicy

Arguments

:: Text
$sel:functionName:GetPolicy'

-> GetPolicy

Create a value of GetPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:qualifier:GetPolicy', getPolicy_qualifier - Specify a version or alias to get the policy for that resource.

$sel:functionName:GetPolicy', getPolicy_functionName - The name of the Lambda function, version, or alias.

Name formats

Function name - my-function (name-only), my-function:v1 (with alias).

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

You can append a version number or alias to any of the formats. The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

Request Lenses

getPolicy_qualifier :: Lens' GetPolicy (Maybe Text)

Specify a version or alias to get the policy for that resource.

getPolicy_functionName :: Lens' GetPolicy Text

The name of the Lambda function, version, or alias.

Name formats

Function name - my-function (name-only), my-function:v1 (with alias).

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

You can append a version number or alias to any of the formats. The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

Destructuring the Response

data GetPolicyResponse

See: newGetPolicyResponse smart constructor.

Constructors

GetPolicyResponse'

Fields

policy :: Maybe Text

The resource-based policy.

revisionId :: Maybe Text

A unique identifier for the current revision of the policy.

httpStatus :: Int

The response's http status code.

newGetPolicyResponse

Arguments

:: Int
$sel:httpStatus:GetPolicyResponse'

-> GetPolicyResponse

Create a value of GetPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policy:GetPolicyResponse', getPolicyResponse_policy - The resource-based policy.

$sel:revisionId:GetPolicyResponse', getPolicyResponse_revisionId - A unique identifier for the current revision of the policy.

$sel:httpStatus:GetPolicyResponse', getPolicyResponse_httpStatus - The response's http status code.

Response Lenses

getPolicyResponse_policy :: Lens' GetPolicyResponse (Maybe Text)

The resource-based policy.

getPolicyResponse_revisionId :: Lens' GetPolicyResponse (Maybe Text)

A unique identifier for the current revision of the policy.

getPolicyResponse_httpStatus :: Lens' GetPolicyResponse Int

The response's http status code.