Network.AWS.Lambda.GetProvisionedConcurrencyConfig

Description

Retrieves the provisioned concurrency configuration for a function's alias or version.

Creating a Request

data GetProvisionedConcurrencyConfig

See: newGetProvisionedConcurrencyConfig smart constructor.

Constructors

GetProvisionedConcurrencyConfig'

Fields

functionName :: Text

The name of the Lambda function.

Name formats

Function name - my-function.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

qualifier :: Text

The version number or alias name.

newGetProvisionedConcurrencyConfig

Arguments

:: Text
$sel:functionName:GetProvisionedConcurrencyConfig'

-> Text
$sel:qualifier:GetProvisionedConcurrencyConfig'

-> GetProvisionedConcurrencyConfig

Create a value of GetProvisionedConcurrencyConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:functionName:GetProvisionedConcurrencyConfig', getProvisionedConcurrencyConfig_functionName - The name of the Lambda function.

Name formats

Function name - my-function.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

$sel:qualifier:GetProvisionedConcurrencyConfig', getProvisionedConcurrencyConfig_qualifier - The version number or alias name.

Request Lenses

getProvisionedConcurrencyConfig_functionName :: Lens' GetProvisionedConcurrencyConfig Text

The name of the Lambda function.

Name formats

Function name - my-function.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

getProvisionedConcurrencyConfig_qualifier :: Lens' GetProvisionedConcurrencyConfig Text

The version number or alias name.

Destructuring the Response

data GetProvisionedConcurrencyConfigResponse

See: newGetProvisionedConcurrencyConfigResponse smart constructor.

Constructors

GetProvisionedConcurrencyConfigResponse'

Fields

status :: Maybe ProvisionedConcurrencyStatusEnum

The status of the allocation process.

availableProvisionedConcurrentExecutions :: Maybe Natural

The amount of provisioned concurrency available.

requestedProvisionedConcurrentExecutions :: Maybe Natural

The amount of provisioned concurrency requested.

allocatedProvisionedConcurrentExecutions :: Maybe Natural

The amount of provisioned concurrency allocated.

lastModified :: Maybe Text

The date and time that a user last updated the configuration, in ISO 8601 format.

statusReason :: Maybe Text

For failed allocations, the reason that provisioned concurrency could not be allocated.

httpStatus :: Int

The response's http status code.

newGetProvisionedConcurrencyConfigResponse

Arguments

:: Int
$sel:httpStatus:GetProvisionedConcurrencyConfigResponse'

-> GetProvisionedConcurrencyConfigResponse

Create a value of GetProvisionedConcurrencyConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetProvisionedConcurrencyConfigResponse', getProvisionedConcurrencyConfigResponse_status - The status of the allocation process.

$sel:availableProvisionedConcurrentExecutions:GetProvisionedConcurrencyConfigResponse', getProvisionedConcurrencyConfigResponse_availableProvisionedConcurrentExecutions - The amount of provisioned concurrency available.

$sel:requestedProvisionedConcurrentExecutions:GetProvisionedConcurrencyConfigResponse', getProvisionedConcurrencyConfigResponse_requestedProvisionedConcurrentExecutions - The amount of provisioned concurrency requested.

$sel:allocatedProvisionedConcurrentExecutions:GetProvisionedConcurrencyConfigResponse', getProvisionedConcurrencyConfigResponse_allocatedProvisionedConcurrentExecutions - The amount of provisioned concurrency allocated.

$sel:lastModified:GetProvisionedConcurrencyConfigResponse', getProvisionedConcurrencyConfigResponse_lastModified - The date and time that a user last updated the configuration, in ISO 8601 format.

$sel:statusReason:GetProvisionedConcurrencyConfigResponse', getProvisionedConcurrencyConfigResponse_statusReason - For failed allocations, the reason that provisioned concurrency could not be allocated.

$sel:httpStatus:GetProvisionedConcurrencyConfigResponse', getProvisionedConcurrencyConfigResponse_httpStatus - The response's http status code.

Response Lenses

getProvisionedConcurrencyConfigResponse_status :: Lens' GetProvisionedConcurrencyConfigResponse (Maybe ProvisionedConcurrencyStatusEnum)

The status of the allocation process.

getProvisionedConcurrencyConfigResponse_availableProvisionedConcurrentExecutions :: Lens' GetProvisionedConcurrencyConfigResponse (Maybe Natural)

The amount of provisioned concurrency available.

getProvisionedConcurrencyConfigResponse_requestedProvisionedConcurrentExecutions :: Lens' GetProvisionedConcurrencyConfigResponse (Maybe Natural)

The amount of provisioned concurrency requested.

getProvisionedConcurrencyConfigResponse_allocatedProvisionedConcurrentExecutions :: Lens' GetProvisionedConcurrencyConfigResponse (Maybe Natural)

The amount of provisioned concurrency allocated.

getProvisionedConcurrencyConfigResponse_lastModified :: Lens' GetProvisionedConcurrencyConfigResponse (Maybe Text)

The date and time that a user last updated the configuration, in ISO 8601 format.

getProvisionedConcurrencyConfigResponse_statusReason :: Lens' GetProvisionedConcurrencyConfigResponse (Maybe Text)

For failed allocations, the reason that provisioned concurrency could not be allocated.

getProvisionedConcurrencyConfigResponse_httpStatus :: Lens' GetProvisionedConcurrencyConfigResponse Int

The response's http status code.