Network.AWS.Lambda.ListEventSourceMappings

Description

Lists event source mappings. Specify an EventSourceArn to only show event source mappings for a single event source.

This operation returns paginated results.

Creating a Request

data ListEventSourceMappings

See: newListEventSourceMappings smart constructor.

Constructors

ListEventSourceMappings'

Fields

eventSourceArn :: Maybe Text

The Amazon Resource Name (ARN) of the event source.

Amazon Kinesis - The ARN of the data stream or a stream consumer.

Amazon DynamoDB Streams - The ARN of the stream.

Amazon Simple Queue Service - The ARN of the queue.

Amazon Managed Streaming for Apache Kafka - The ARN of the cluster.

functionName :: Maybe Text

The name of the Lambda function.

Name formats

Function name - MyFunction.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction.

Version or Alias ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction:PROD.

Partial ARN - 123456789012:function:MyFunction.

The length constraint applies only to the full ARN. If you specify only the function name, it's limited to 64 characters in length.

maxItems :: Maybe Natural

The maximum number of event source mappings to return. Note that ListEventSourceMappings returns a maximum of 100 items in each response, even if you set the number higher.

marker :: Maybe Text

A pagination token returned by a previous call.

newListEventSourceMappings :: ListEventSourceMappings

Create a value of ListEventSourceMappings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventSourceArn:ListEventSourceMappings', listEventSourceMappings_eventSourceArn - The Amazon Resource Name (ARN) of the event source.

Amazon Kinesis - The ARN of the data stream or a stream consumer.

Amazon DynamoDB Streams - The ARN of the stream.

Amazon Simple Queue Service - The ARN of the queue.

Amazon Managed Streaming for Apache Kafka - The ARN of the cluster.

$sel:functionName:ListEventSourceMappings', listEventSourceMappings_functionName - The name of the Lambda function.

Name formats

Function name - MyFunction.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction.

Version or Alias ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction:PROD.

Partial ARN - 123456789012:function:MyFunction.

The length constraint applies only to the full ARN. If you specify only the function name, it's limited to 64 characters in length.

$sel:maxItems:ListEventSourceMappings', listEventSourceMappings_maxItems - The maximum number of event source mappings to return. Note that ListEventSourceMappings returns a maximum of 100 items in each response, even if you set the number higher.

$sel:marker:ListEventSourceMappings', listEventSourceMappings_marker - A pagination token returned by a previous call.

Request Lenses

listEventSourceMappings_eventSourceArn :: Lens' ListEventSourceMappings (Maybe Text)

The Amazon Resource Name (ARN) of the event source.

Amazon Kinesis - The ARN of the data stream or a stream consumer.

Amazon DynamoDB Streams - The ARN of the stream.

Amazon Simple Queue Service - The ARN of the queue.

Amazon Managed Streaming for Apache Kafka - The ARN of the cluster.

listEventSourceMappings_functionName :: Lens' ListEventSourceMappings (Maybe Text)

The name of the Lambda function.

Name formats

Function name - MyFunction.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction.

Version or Alias ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction:PROD.

Partial ARN - 123456789012:function:MyFunction.

The length constraint applies only to the full ARN. If you specify only the function name, it's limited to 64 characters in length.

listEventSourceMappings_maxItems :: Lens' ListEventSourceMappings (Maybe Natural)

The maximum number of event source mappings to return. Note that ListEventSourceMappings returns a maximum of 100 items in each response, even if you set the number higher.

listEventSourceMappings_marker :: Lens' ListEventSourceMappings (Maybe Text)

A pagination token returned by a previous call.

Destructuring the Response

data ListEventSourceMappingsResponse

See: newListEventSourceMappingsResponse smart constructor.

Constructors

ListEventSourceMappingsResponse'

Fields

eventSourceMappings :: Maybe [EventSourceMappingConfiguration]

A list of event source mappings.

nextMarker :: Maybe Text

A pagination token that's returned when the response doesn't contain all event source mappings.

httpStatus :: Int

The response's http status code.

newListEventSourceMappingsResponse

Arguments

:: Int
$sel:httpStatus:ListEventSourceMappingsResponse'

-> ListEventSourceMappingsResponse

Create a value of ListEventSourceMappingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventSourceMappings:ListEventSourceMappingsResponse', listEventSourceMappingsResponse_eventSourceMappings - A list of event source mappings.

$sel:nextMarker:ListEventSourceMappingsResponse', listEventSourceMappingsResponse_nextMarker - A pagination token that's returned when the response doesn't contain all event source mappings.

$sel:httpStatus:ListEventSourceMappingsResponse', listEventSourceMappingsResponse_httpStatus - The response's http status code.

Response Lenses

listEventSourceMappingsResponse_eventSourceMappings :: Lens' ListEventSourceMappingsResponse (Maybe [EventSourceMappingConfiguration])

A list of event source mappings.

listEventSourceMappingsResponse_nextMarker :: Lens' ListEventSourceMappingsResponse (Maybe Text)

A pagination token that's returned when the response doesn't contain all event source mappings.

listEventSourceMappingsResponse_httpStatus :: Lens' ListEventSourceMappingsResponse Int

The response's http status code.