Network.AWS.Lambda.ListLayerVersions

Description

Lists the versions of an Lambda layer. Versions that have been deleted aren't listed. Specify a runtime identifier to list only versions that indicate that they're compatible with that runtime. Specify a compatible architecture to include only layer versions that are compatible with that architecture.

This operation returns paginated results.

Creating a Request

data ListLayerVersions

See: newListLayerVersions smart constructor.

Constructors

ListLayerVersions'

Fields

compatibleRuntime :: Maybe Runtime

A runtime identifier. For example, go1.x.

marker :: Maybe Text

A pagination token returned by a previous call.

maxItems :: Maybe Natural

The maximum number of versions to return.

compatibleArchitecture :: Maybe Architecture

The compatible instruction set architecture.

layerName :: Text

The name or Amazon Resource Name (ARN) of the layer.

newListLayerVersions

Arguments

:: Text
$sel:layerName:ListLayerVersions'

-> ListLayerVersions

Create a value of ListLayerVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:compatibleRuntime:ListLayerVersions', listLayerVersions_compatibleRuntime - A runtime identifier. For example, go1.x.

$sel:marker:ListLayerVersions', listLayerVersions_marker - A pagination token returned by a previous call.

$sel:maxItems:ListLayerVersions', listLayerVersions_maxItems - The maximum number of versions to return.

$sel:compatibleArchitecture:ListLayerVersions', listLayerVersions_compatibleArchitecture - The compatible instruction set architecture.

$sel:layerName:ListLayerVersions', listLayerVersions_layerName - The name or Amazon Resource Name (ARN) of the layer.

Request Lenses

listLayerVersions_compatibleRuntime :: Lens' ListLayerVersions (Maybe Runtime)

A runtime identifier. For example, go1.x.

listLayerVersions_marker :: Lens' ListLayerVersions (Maybe Text)

A pagination token returned by a previous call.

listLayerVersions_maxItems :: Lens' ListLayerVersions (Maybe Natural)

The maximum number of versions to return.

listLayerVersions_compatibleArchitecture :: Lens' ListLayerVersions (Maybe Architecture)

The compatible instruction set architecture.

listLayerVersions_layerName :: Lens' ListLayerVersions Text

The name or Amazon Resource Name (ARN) of the layer.

Destructuring the Response

data ListLayerVersionsResponse

See: newListLayerVersionsResponse smart constructor.

Constructors

ListLayerVersionsResponse'

Fields

layerVersions :: Maybe [LayerVersionsListItem]

A list of versions.

nextMarker :: Maybe Text

A pagination token returned when the response doesn't contain all versions.

httpStatus :: Int

The response's http status code.

newListLayerVersionsResponse

Arguments

:: Int
$sel:httpStatus:ListLayerVersionsResponse'

-> ListLayerVersionsResponse

Create a value of ListLayerVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:layerVersions:ListLayerVersionsResponse', listLayerVersionsResponse_layerVersions - A list of versions.

$sel:nextMarker:ListLayerVersionsResponse', listLayerVersionsResponse_nextMarker - A pagination token returned when the response doesn't contain all versions.

$sel:httpStatus:ListLayerVersionsResponse', listLayerVersionsResponse_httpStatus - The response's http status code.

Response Lenses

listLayerVersionsResponse_layerVersions :: Lens' ListLayerVersionsResponse (Maybe [LayerVersionsListItem])

A list of versions.

listLayerVersionsResponse_nextMarker :: Lens' ListLayerVersionsResponse (Maybe Text)

A pagination token returned when the response doesn't contain all versions.

listLayerVersionsResponse_httpStatus :: Lens' ListLayerVersionsResponse Int

The response's http status code.