Network.AWS.Lambda.ListLayers

Description

Lists Lambda layers and shows information about the latest version of each. Specify a runtime identifier to list only layers that indicate that they're compatible with that runtime.

This operation returns paginated results.

Creating a Request

data ListLayers

See: newListLayers smart constructor.

Constructors

ListLayers'

Fields

maxItems :: Maybe Natural

The maximum number of layers to return.

marker :: Maybe Text

A pagination token returned by a previous call.

compatibleRuntime :: Maybe Runtime

A runtime identifier. For example, go1.x.

newListLayers :: ListLayers

Create a value of ListLayers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxItems:ListLayers', listLayers_maxItems - The maximum number of layers to return.

$sel:marker:ListLayers', listLayers_marker - A pagination token returned by a previous call.

$sel:compatibleRuntime:ListLayers', listLayers_compatibleRuntime - A runtime identifier. For example, go1.x.

Request Lenses

listLayers_maxItems :: Lens' ListLayers (Maybe Natural)

The maximum number of layers to return.

listLayers_marker :: Lens' ListLayers (Maybe Text)

A pagination token returned by a previous call.

listLayers_compatibleRuntime :: Lens' ListLayers (Maybe Runtime)

A runtime identifier. For example, go1.x.

Destructuring the Response

data ListLayersResponse

See: newListLayersResponse smart constructor.

Constructors

ListLayersResponse'

Fields

nextMarker :: Maybe Text

A pagination token returned when the response doesn't contain all layers.

layers :: Maybe [LayersListItem]

A list of function layers.

httpStatus :: Int

The response's http status code.

newListLayersResponse

Arguments

:: Int
$sel:httpStatus:ListLayersResponse'

-> ListLayersResponse

Create a value of ListLayersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextMarker:ListLayersResponse', listLayersResponse_nextMarker - A pagination token returned when the response doesn't contain all layers.

$sel:layers:ListLayersResponse', listLayersResponse_layers - A list of function layers.

$sel:httpStatus:ListLayersResponse', listLayersResponse_httpStatus - The response's http status code.

Response Lenses

listLayersResponse_nextMarker :: Lens' ListLayersResponse (Maybe Text)

A pagination token returned when the response doesn't contain all layers.

listLayersResponse_layers :: Lens' ListLayersResponse (Maybe [LayersListItem])

A list of function layers.

listLayersResponse_httpStatus :: Lens' ListLayersResponse Int

The response's http status code.