Network.AWS.Lambda.PublishLayerVersion

Description

Creates an Lambda layer from a ZIP archive. Each time you call PublishLayerVersion with the same layer name, a new version is created.

Add layers to your function with CreateFunction or UpdateFunctionConfiguration.

Creating a Request

data PublishLayerVersion

See: newPublishLayerVersion smart constructor.

Constructors

PublishLayerVersion'

Fields

compatibleRuntimes :: Maybe [Runtime]

A list of compatible function runtimes. Used for filtering with ListLayers and ListLayerVersions.

description :: Maybe Text

The description of the version.

licenseInfo :: Maybe Text

The layer's software license. It can be any of the following:

An SPDX license identifier. For example, MIT.

The URL of a license hosted on the internet. For example, https://opensource.org/licenses/MIT.

The full text of the license.

layerName :: Text

The name or Amazon Resource Name (ARN) of the layer.

content :: LayerVersionContentInput

The function layer archive.

newPublishLayerVersion

Arguments

:: Text
$sel:layerName:PublishLayerVersion'

-> LayerVersionContentInput
$sel:content:PublishLayerVersion'

-> PublishLayerVersion

Create a value of PublishLayerVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:compatibleRuntimes:PublishLayerVersion', publishLayerVersion_compatibleRuntimes - A list of compatible function runtimes. Used for filtering with ListLayers and ListLayerVersions.

$sel:description:PublishLayerVersion', publishLayerVersion_description - The description of the version.

$sel:licenseInfo:PublishLayerVersion', publishLayerVersion_licenseInfo - The layer's software license. It can be any of the following:

An SPDX license identifier. For example, MIT.

The URL of a license hosted on the internet. For example, https://opensource.org/licenses/MIT.

The full text of the license.

$sel:layerName:PublishLayerVersion', publishLayerVersion_layerName - The name or Amazon Resource Name (ARN) of the layer.

$sel:content:PublishLayerVersion', publishLayerVersion_content - The function layer archive.

Request Lenses

publishLayerVersion_compatibleRuntimes :: Lens' PublishLayerVersion (Maybe [Runtime])

A list of compatible function runtimes. Used for filtering with ListLayers and ListLayerVersions.

publishLayerVersion_description :: Lens' PublishLayerVersion (Maybe Text)

The description of the version.

publishLayerVersion_licenseInfo :: Lens' PublishLayerVersion (Maybe Text)

The layer's software license. It can be any of the following:

An SPDX license identifier. For example, MIT.

The URL of a license hosted on the internet. For example, https://opensource.org/licenses/MIT.

The full text of the license.

publishLayerVersion_layerName :: Lens' PublishLayerVersion Text

The name or Amazon Resource Name (ARN) of the layer.

publishLayerVersion_content :: Lens' PublishLayerVersion LayerVersionContentInput

The function layer archive.

Destructuring the Response

data PublishLayerVersionResponse

See: newPublishLayerVersionResponse smart constructor.

Constructors

PublishLayerVersionResponse'

Fields

createdDate :: Maybe Text

The date that the layer version was created, in ISO-8601 format (YYYY-MM-DDThh:mm:ss.sTZD).

layerArn :: Maybe Text

The ARN of the layer.

version :: Maybe Integer

The version number.

layerVersionArn :: Maybe Text

The ARN of the layer version.

content :: Maybe LayerVersionContentOutput

Details about the layer version.

compatibleRuntimes :: Maybe [Runtime]

The layer's compatible runtimes.

description :: Maybe Text

The description of the version.

licenseInfo :: Maybe Text

The layer's software license.

httpStatus :: Int

The response's http status code.

newPublishLayerVersionResponse

Arguments

:: Int
$sel:httpStatus:PublishLayerVersionResponse'

-> PublishLayerVersionResponse

Create a value of PublishLayerVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdDate:PublishLayerVersionResponse', publishLayerVersionResponse_createdDate - The date that the layer version was created, in ISO-8601 format (YYYY-MM-DDThh:mm:ss.sTZD).

$sel:layerArn:PublishLayerVersionResponse', publishLayerVersionResponse_layerArn - The ARN of the layer.

$sel:version:PublishLayerVersionResponse', publishLayerVersionResponse_version - The version number.

$sel:layerVersionArn:PublishLayerVersionResponse', publishLayerVersionResponse_layerVersionArn - The ARN of the layer version.

$sel:content:PublishLayerVersion', publishLayerVersionResponse_content - Details about the layer version.

$sel:compatibleRuntimes:PublishLayerVersion', publishLayerVersionResponse_compatibleRuntimes - The layer's compatible runtimes.

$sel:description:PublishLayerVersion', publishLayerVersionResponse_description - The description of the version.

$sel:licenseInfo:PublishLayerVersion', publishLayerVersionResponse_licenseInfo - The layer's software license.

$sel:httpStatus:PublishLayerVersionResponse', publishLayerVersionResponse_httpStatus - The response's http status code.

Response Lenses

publishLayerVersionResponse_createdDate :: Lens' PublishLayerVersionResponse (Maybe Text)

The date that the layer version was created, in ISO-8601 format (YYYY-MM-DDThh:mm:ss.sTZD).

publishLayerVersionResponse_layerArn :: Lens' PublishLayerVersionResponse (Maybe Text)

The ARN of the layer.

publishLayerVersionResponse_version :: Lens' PublishLayerVersionResponse (Maybe Integer)

The version number.

publishLayerVersionResponse_layerVersionArn :: Lens' PublishLayerVersionResponse (Maybe Text)

The ARN of the layer version.

publishLayerVersionResponse_content :: Lens' PublishLayerVersionResponse (Maybe LayerVersionContentOutput)

Details about the layer version.

publishLayerVersionResponse_compatibleRuntimes :: Lens' PublishLayerVersionResponse (Maybe [Runtime])

The layer's compatible runtimes.

publishLayerVersionResponse_description :: Lens' PublishLayerVersionResponse (Maybe Text)

The description of the version.

publishLayerVersionResponse_licenseInfo :: Lens' PublishLayerVersionResponse (Maybe Text)

The layer's software license.

publishLayerVersionResponse_httpStatus :: Lens' PublishLayerVersionResponse Int

The response's http status code.