Network.AWS.Lambda.PublishVersion

Description

Creates a version from the current code and configuration of a function. Use versions to create a snapshot of your function code and configuration that doesn't change.

Lambda doesn't publish a version if the function's configuration and code haven't changed since the last version. Use UpdateFunctionCode or UpdateFunctionConfiguration to update the function before publishing a version.

Clients can invoke versions directly or with an alias. To create an alias, use CreateAlias.

Creating a Request

data PublishVersion

See: newPublishVersion smart constructor.

Constructors

PublishVersion'

Fields

codeSha256 :: Maybe Text

Only publish a version if the hash value matches the value that's specified. Use this option to avoid publishing a version if the function code has changed since you last updated it. You can get the hash for the version that you uploaded from the output of UpdateFunctionCode.

description :: Maybe Text

A description for the version to override the description in the function configuration.

revisionId :: Maybe Text

Only update the function if the revision ID matches the ID that's specified. Use this option to avoid publishing a version if the function configuration has changed since you last updated it.

functionName :: Text

The name of the Lambda function.

Name formats

Function name - MyFunction.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction.

Partial ARN - 123456789012:function:MyFunction.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

newPublishVersion

Arguments

:: Text
$sel:functionName:PublishVersion'

-> PublishVersion

Create a value of PublishVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:codeSha256:PublishVersion', publishVersion_codeSha256 - Only publish a version if the hash value matches the value that's specified. Use this option to avoid publishing a version if the function code has changed since you last updated it. You can get the hash for the version that you uploaded from the output of UpdateFunctionCode.

$sel:description:PublishVersion', publishVersion_description - A description for the version to override the description in the function configuration.

$sel:revisionId:PublishVersion', publishVersion_revisionId - Only update the function if the revision ID matches the ID that's specified. Use this option to avoid publishing a version if the function configuration has changed since you last updated it.

$sel:functionName:PublishVersion', publishVersion_functionName - The name of the Lambda function.

Name formats

Function name - MyFunction.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction.

Partial ARN - 123456789012:function:MyFunction.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

Request Lenses

publishVersion_codeSha256 :: Lens' PublishVersion (Maybe Text)

Only publish a version if the hash value matches the value that's specified. Use this option to avoid publishing a version if the function code has changed since you last updated it. You can get the hash for the version that you uploaded from the output of UpdateFunctionCode.

publishVersion_description :: Lens' PublishVersion (Maybe Text)

A description for the version to override the description in the function configuration.

publishVersion_revisionId :: Lens' PublishVersion (Maybe Text)

Only update the function if the revision ID matches the ID that's specified. Use this option to avoid publishing a version if the function configuration has changed since you last updated it.

publishVersion_functionName :: Lens' PublishVersion Text

The name of the Lambda function.

Name formats

Function name - MyFunction.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction.

Partial ARN - 123456789012:function:MyFunction.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

Destructuring the Response

data FunctionConfiguration

Details about a function's configuration.

See: newFunctionConfiguration smart constructor.

Constructors

FunctionConfiguration'

Fields

memorySize :: Maybe Natural

The amount of memory available to the function at runtime.

runtime :: Maybe Runtime

The runtime environment for the Lambda function.

state :: Maybe State

The current state of the function. When the state is Inactive, you can reactivate the function by invoking it.

signingProfileVersionArn :: Maybe Text

The ARN of the signing profile version.

lastUpdateStatus :: Maybe LastUpdateStatus

The status of the last update that was performed on the function. This is first set to Successful after function creation completes.

functionArn :: Maybe Text

The function's Amazon Resource Name (ARN).

kmsKeyArn :: Maybe Text

The KMS key that's used to encrypt the function's environment variables. This key is only returned if you've configured a customer managed CMK.

packageType :: Maybe PackageType

The type of deployment package. Set to Image for container image and set Zip for .zip file archive.

fileSystemConfigs :: Maybe [FileSystemConfig]

Connection settings for an Amazon EFS file system.

environment :: Maybe EnvironmentResponse

The function's environment variables.

deadLetterConfig :: Maybe DeadLetterConfig

The function's dead letter queue.

architectures :: Maybe (NonEmpty Architecture)

The instruction set architecture that the function supports. Architecture is a string array with one of the valid values. The default architecture value is x86_64.

signingJobArn :: Maybe Text

The ARN of the signing job.

role' :: Maybe Text

The function's execution role.

vpcConfig :: Maybe VpcConfigResponse

The function's networking configuration.

version :: Maybe Text

The version of the Lambda function.

functionName :: Maybe Text

The name of the function.

layers :: Maybe [Layer]

The function's layers.

codeSize :: Maybe Integer

The size of the function's deployment package, in bytes.

handler :: Maybe Text

The function that Lambda calls to begin executing your function.

timeout :: Maybe Natural

The amount of time in seconds that Lambda allows a function to run before stopping it.

lastUpdateStatusReason :: Maybe Text

The reason for the last update that was performed on the function.

stateReason :: Maybe Text

The reason for the function's current state.

lastModified :: Maybe Text

The date and time that the function was last updated, in ISO-8601 format (YYYY-MM-DDThh:mm:ss.sTZD).

codeSha256 :: Maybe Text

The SHA256 hash of the function's deployment package.

tracingConfig :: Maybe TracingConfigResponse

The function's X-Ray tracing configuration.

stateReasonCode :: Maybe StateReasonCode

The reason code for the function's current state. When the code is Creating, you can't invoke or modify the function.

imageConfigResponse :: Maybe ImageConfigResponse

The function's image configuration values.

description :: Maybe Text

The function's description.

lastUpdateStatusReasonCode :: Maybe LastUpdateStatusReasonCode

The reason code for the last update that was performed on the function.

revisionId :: Maybe Text

The latest updated revision of the function or alias.

masterArn :: Maybe Text

For Lambda@Edge functions, the ARN of the master function.

newFunctionConfiguration :: FunctionConfiguration

Create a value of FunctionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memorySize:FunctionConfiguration', functionConfiguration_memorySize - The amount of memory available to the function at runtime.

$sel:runtime:FunctionConfiguration', functionConfiguration_runtime - The runtime environment for the Lambda function.

$sel:state:FunctionConfiguration', functionConfiguration_state - The current state of the function. When the state is Inactive, you can reactivate the function by invoking it.

$sel:signingProfileVersionArn:FunctionConfiguration', functionConfiguration_signingProfileVersionArn - The ARN of the signing profile version.

$sel:lastUpdateStatus:FunctionConfiguration', functionConfiguration_lastUpdateStatus - The status of the last update that was performed on the function. This is first set to Successful after function creation completes.

$sel:functionArn:FunctionConfiguration', functionConfiguration_functionArn - The function's Amazon Resource Name (ARN).

$sel:kmsKeyArn:FunctionConfiguration', functionConfiguration_kmsKeyArn - The KMS key that's used to encrypt the function's environment variables. This key is only returned if you've configured a customer managed CMK.

$sel:packageType:FunctionConfiguration', functionConfiguration_packageType - The type of deployment package. Set to Image for container image and set Zip for .zip file archive.

$sel:fileSystemConfigs:FunctionConfiguration', functionConfiguration_fileSystemConfigs - Connection settings for an Amazon EFS file system.

$sel:environment:FunctionConfiguration', functionConfiguration_environment - The function's environment variables.

$sel:deadLetterConfig:FunctionConfiguration', functionConfiguration_deadLetterConfig - The function's dead letter queue.

$sel:architectures:FunctionConfiguration', functionConfiguration_architectures - The instruction set architecture that the function supports. Architecture is a string array with one of the valid values. The default architecture value is x86_64.

$sel:signingJobArn:FunctionConfiguration', functionConfiguration_signingJobArn - The ARN of the signing job.

$sel:role':FunctionConfiguration', functionConfiguration_role - The function's execution role.

$sel:vpcConfig:FunctionConfiguration', functionConfiguration_vpcConfig - The function's networking configuration.

$sel:version:FunctionConfiguration', functionConfiguration_version - The version of the Lambda function.

$sel:functionName:FunctionConfiguration', functionConfiguration_functionName - The name of the function.

$sel:layers:FunctionConfiguration', functionConfiguration_layers - The function's layers.

$sel:codeSize:FunctionConfiguration', functionConfiguration_codeSize - The size of the function's deployment package, in bytes.

$sel:handler:FunctionConfiguration', functionConfiguration_handler - The function that Lambda calls to begin executing your function.

$sel:timeout:FunctionConfiguration', functionConfiguration_timeout - The amount of time in seconds that Lambda allows a function to run before stopping it.

$sel:lastUpdateStatusReason:FunctionConfiguration', functionConfiguration_lastUpdateStatusReason - The reason for the last update that was performed on the function.

$sel:stateReason:FunctionConfiguration', functionConfiguration_stateReason - The reason for the function's current state.

$sel:lastModified:FunctionConfiguration', functionConfiguration_lastModified - The date and time that the function was last updated, in ISO-8601 format (YYYY-MM-DDThh:mm:ss.sTZD).

$sel:codeSha256:FunctionConfiguration', functionConfiguration_codeSha256 - The SHA256 hash of the function's deployment package.

$sel:tracingConfig:FunctionConfiguration', functionConfiguration_tracingConfig - The function's X-Ray tracing configuration.

$sel:stateReasonCode:FunctionConfiguration', functionConfiguration_stateReasonCode - The reason code for the function's current state. When the code is Creating, you can't invoke or modify the function.

$sel:imageConfigResponse:FunctionConfiguration', functionConfiguration_imageConfigResponse - The function's image configuration values.

$sel:description:FunctionConfiguration', functionConfiguration_description - The function's description.

$sel:lastUpdateStatusReasonCode:FunctionConfiguration', functionConfiguration_lastUpdateStatusReasonCode - The reason code for the last update that was performed on the function.

$sel:revisionId:FunctionConfiguration', functionConfiguration_revisionId - The latest updated revision of the function or alias.

$sel:masterArn:FunctionConfiguration', functionConfiguration_masterArn - For Lambda@Edge functions, the ARN of the master function.

Response Lenses

functionConfiguration_memorySize :: Lens' FunctionConfiguration (Maybe Natural)

The amount of memory available to the function at runtime.

functionConfiguration_runtime :: Lens' FunctionConfiguration (Maybe Runtime)

The runtime environment for the Lambda function.

functionConfiguration_state :: Lens' FunctionConfiguration (Maybe State)

The current state of the function. When the state is Inactive, you can reactivate the function by invoking it.

functionConfiguration_signingProfileVersionArn :: Lens' FunctionConfiguration (Maybe Text)

The ARN of the signing profile version.

functionConfiguration_lastUpdateStatus :: Lens' FunctionConfiguration (Maybe LastUpdateStatus)

The status of the last update that was performed on the function. This is first set to Successful after function creation completes.

functionConfiguration_functionArn :: Lens' FunctionConfiguration (Maybe Text)

The function's Amazon Resource Name (ARN).

functionConfiguration_kmsKeyArn :: Lens' FunctionConfiguration (Maybe Text)

The KMS key that's used to encrypt the function's environment variables. This key is only returned if you've configured a customer managed CMK.

functionConfiguration_packageType :: Lens' FunctionConfiguration (Maybe PackageType)

The type of deployment package. Set to Image for container image and set Zip for .zip file archive.

functionConfiguration_fileSystemConfigs :: Lens' FunctionConfiguration (Maybe [FileSystemConfig])

Connection settings for an Amazon EFS file system.

functionConfiguration_environment :: Lens' FunctionConfiguration (Maybe EnvironmentResponse)

The function's environment variables.

functionConfiguration_deadLetterConfig :: Lens' FunctionConfiguration (Maybe DeadLetterConfig)

The function's dead letter queue.

functionConfiguration_architectures :: Lens' FunctionConfiguration (Maybe (NonEmpty Architecture))

The instruction set architecture that the function supports. Architecture is a string array with one of the valid values. The default architecture value is x86_64.

functionConfiguration_signingJobArn :: Lens' FunctionConfiguration (Maybe Text)

The ARN of the signing job.

functionConfiguration_role :: Lens' FunctionConfiguration (Maybe Text)

The function's execution role.

functionConfiguration_vpcConfig :: Lens' FunctionConfiguration (Maybe VpcConfigResponse)

The function's networking configuration.

functionConfiguration_version :: Lens' FunctionConfiguration (Maybe Text)

The version of the Lambda function.

functionConfiguration_functionName :: Lens' FunctionConfiguration (Maybe Text)

The name of the function.

functionConfiguration_layers :: Lens' FunctionConfiguration (Maybe [Layer])

The function's layers.

functionConfiguration_codeSize :: Lens' FunctionConfiguration (Maybe Integer)

The size of the function's deployment package, in bytes.

functionConfiguration_handler :: Lens' FunctionConfiguration (Maybe Text)

The function that Lambda calls to begin executing your function.

functionConfiguration_timeout :: Lens' FunctionConfiguration (Maybe Natural)

The amount of time in seconds that Lambda allows a function to run before stopping it.

functionConfiguration_lastUpdateStatusReason :: Lens' FunctionConfiguration (Maybe Text)

The reason for the last update that was performed on the function.

functionConfiguration_stateReason :: Lens' FunctionConfiguration (Maybe Text)

The reason for the function's current state.

functionConfiguration_lastModified :: Lens' FunctionConfiguration (Maybe Text)

The date and time that the function was last updated, in ISO-8601 format (YYYY-MM-DDThh:mm:ss.sTZD).

functionConfiguration_codeSha256 :: Lens' FunctionConfiguration (Maybe Text)

The SHA256 hash of the function's deployment package.

functionConfiguration_tracingConfig :: Lens' FunctionConfiguration (Maybe TracingConfigResponse)

The function's X-Ray tracing configuration.

functionConfiguration_stateReasonCode :: Lens' FunctionConfiguration (Maybe StateReasonCode)

The reason code for the function's current state. When the code is Creating, you can't invoke or modify the function.

functionConfiguration_imageConfigResponse :: Lens' FunctionConfiguration (Maybe ImageConfigResponse)

The function's image configuration values.

functionConfiguration_description :: Lens' FunctionConfiguration (Maybe Text)

The function's description.

functionConfiguration_lastUpdateStatusReasonCode :: Lens' FunctionConfiguration (Maybe LastUpdateStatusReasonCode)

The reason code for the last update that was performed on the function.

functionConfiguration_revisionId :: Lens' FunctionConfiguration (Maybe Text)

The latest updated revision of the function or alias.

functionConfiguration_masterArn :: Lens' FunctionConfiguration (Maybe Text)

For Lambda@Edge functions, the ARN of the master function.