Network.AWS.Lambda.PutFunctionCodeSigningConfig

Description

Update the code signing configuration for the function. Changes to the code signing configuration take effect the next time a user tries to deploy a code package to the function.

Creating a Request

data PutFunctionCodeSigningConfig

See: newPutFunctionCodeSigningConfig smart constructor.

Constructors

PutFunctionCodeSigningConfig'

Fields

codeSigningConfigArn :: Text

The The Amazon Resource Name (ARN) of the code signing configuration.

functionName :: Text

The name of the Lambda function.

Name formats

Function name - MyFunction.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction.

Partial ARN - 123456789012:function:MyFunction.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

newPutFunctionCodeSigningConfig

Arguments

:: Text
$sel:codeSigningConfigArn:PutFunctionCodeSigningConfig'

-> Text
$sel:functionName:PutFunctionCodeSigningConfig'

-> PutFunctionCodeSigningConfig

Create a value of PutFunctionCodeSigningConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:codeSigningConfigArn:PutFunctionCodeSigningConfig', putFunctionCodeSigningConfig_codeSigningConfigArn - The The Amazon Resource Name (ARN) of the code signing configuration.

$sel:functionName:PutFunctionCodeSigningConfig', putFunctionCodeSigningConfig_functionName - The name of the Lambda function.

Name formats

Function name - MyFunction.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction.

Partial ARN - 123456789012:function:MyFunction.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

Request Lenses

putFunctionCodeSigningConfig_codeSigningConfigArn :: Lens' PutFunctionCodeSigningConfig Text

The The Amazon Resource Name (ARN) of the code signing configuration.

putFunctionCodeSigningConfig_functionName :: Lens' PutFunctionCodeSigningConfig Text

The name of the Lambda function.

Name formats

Function name - MyFunction.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction.

Partial ARN - 123456789012:function:MyFunction.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

Destructuring the Response

data PutFunctionCodeSigningConfigResponse

See: newPutFunctionCodeSigningConfigResponse smart constructor.

Constructors

PutFunctionCodeSigningConfigResponse'

Fields

httpStatus :: Int

The response's http status code.

codeSigningConfigArn :: Text

The The Amazon Resource Name (ARN) of the code signing configuration.

functionName :: Text

The name of the Lambda function.

Name formats

Function name - MyFunction.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction.

Partial ARN - 123456789012:function:MyFunction.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

newPutFunctionCodeSigningConfigResponse

Arguments

:: Int
$sel:httpStatus:PutFunctionCodeSigningConfigResponse'

-> Text
$sel:codeSigningConfigArn:PutFunctionCodeSigningConfig'

-> Text
$sel:functionName:PutFunctionCodeSigningConfig'

-> PutFunctionCodeSigningConfigResponse

Create a value of PutFunctionCodeSigningConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutFunctionCodeSigningConfigResponse', putFunctionCodeSigningConfigResponse_httpStatus - The response's http status code.

$sel:codeSigningConfigArn:PutFunctionCodeSigningConfig', putFunctionCodeSigningConfigResponse_codeSigningConfigArn - The The Amazon Resource Name (ARN) of the code signing configuration.

$sel:functionName:PutFunctionCodeSigningConfig', putFunctionCodeSigningConfigResponse_functionName - The name of the Lambda function.

Name formats

Function name - MyFunction.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction.

Partial ARN - 123456789012:function:MyFunction.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

Response Lenses

putFunctionCodeSigningConfigResponse_httpStatus :: Lens' PutFunctionCodeSigningConfigResponse Int

The response's http status code.

putFunctionCodeSigningConfigResponse_codeSigningConfigArn :: Lens' PutFunctionCodeSigningConfigResponse Text

The The Amazon Resource Name (ARN) of the code signing configuration.

putFunctionCodeSigningConfigResponse_functionName :: Lens' PutFunctionCodeSigningConfigResponse Text

The name of the Lambda function.

Name formats

Function name - MyFunction.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction.

Partial ARN - 123456789012:function:MyFunction.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.