Network.AWS.Lambda.PutFunctionEventInvokeConfig

Description

Configures options for asynchronous invocation on a function, version, or alias. If a configuration already exists for a function, version, or alias, this operation overwrites it. If you exclude any settings, they are removed. To set one option without affecting existing settings for other options, use UpdateFunctionEventInvokeConfig.

By default, Lambda retries an asynchronous invocation twice if the function returns an error. It retains events in a queue for up to six hours. When an event fails all processing attempts or stays in the asynchronous invocation queue for too long, Lambda discards it. To retain discarded events, configure a dead-letter queue with UpdateFunctionConfiguration.

To send an invocation record to a queue, topic, function, or event bus, specify a destination. You can configure separate destinations for successful invocations (on-success) and events that fail all processing attempts (on-failure). You can configure destinations in addition to or instead of a dead-letter queue.

Creating a Request

data PutFunctionEventInvokeConfig

See: newPutFunctionEventInvokeConfig smart constructor.

Constructors

PutFunctionEventInvokeConfig'

Fields

maximumEventAgeInSeconds :: Maybe Natural

The maximum age of a request that Lambda sends to a function for processing.

qualifier :: Maybe Text

A version number or alias name.

destinationConfig :: Maybe DestinationConfig

A destination for events after they have been sent to a function for processing.

Destinations

Function - The Amazon Resource Name (ARN) of a Lambda function.

Queue - The ARN of an SQS queue.

Topic - The ARN of an SNS topic.

Event Bus - The ARN of an Amazon EventBridge event bus.

maximumRetryAttempts :: Maybe Natural

The maximum number of times to retry when the function returns an error.

functionName :: Text

The name of the Lambda function, version, or alias.

Name formats

Function name - my-function (name-only), my-function:v1 (with alias).

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

You can append a version number or alias to any of the formats. The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

newPutFunctionEventInvokeConfig

Arguments

:: Text
$sel:functionName:PutFunctionEventInvokeConfig'

-> PutFunctionEventInvokeConfig

Create a value of PutFunctionEventInvokeConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maximumEventAgeInSeconds:PutFunctionEventInvokeConfig', putFunctionEventInvokeConfig_maximumEventAgeInSeconds - The maximum age of a request that Lambda sends to a function for processing.

$sel:qualifier:PutFunctionEventInvokeConfig', putFunctionEventInvokeConfig_qualifier - A version number or alias name.

$sel:destinationConfig:PutFunctionEventInvokeConfig', putFunctionEventInvokeConfig_destinationConfig - A destination for events after they have been sent to a function for processing.

Destinations

Function - The Amazon Resource Name (ARN) of a Lambda function.

Queue - The ARN of an SQS queue.

Topic - The ARN of an SNS topic.

Event Bus - The ARN of an Amazon EventBridge event bus.

$sel:maximumRetryAttempts:PutFunctionEventInvokeConfig', putFunctionEventInvokeConfig_maximumRetryAttempts - The maximum number of times to retry when the function returns an error.

$sel:functionName:PutFunctionEventInvokeConfig', putFunctionEventInvokeConfig_functionName - The name of the Lambda function, version, or alias.

Name formats

Function name - my-function (name-only), my-function:v1 (with alias).

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

You can append a version number or alias to any of the formats. The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

Request Lenses

putFunctionEventInvokeConfig_maximumEventAgeInSeconds :: Lens' PutFunctionEventInvokeConfig (Maybe Natural)

The maximum age of a request that Lambda sends to a function for processing.

putFunctionEventInvokeConfig_qualifier :: Lens' PutFunctionEventInvokeConfig (Maybe Text)

A version number or alias name.

putFunctionEventInvokeConfig_destinationConfig :: Lens' PutFunctionEventInvokeConfig (Maybe DestinationConfig)

A destination for events after they have been sent to a function for processing.

Destinations

Function - The Amazon Resource Name (ARN) of a Lambda function.

Queue - The ARN of an SQS queue.

Topic - The ARN of an SNS topic.

Event Bus - The ARN of an Amazon EventBridge event bus.

putFunctionEventInvokeConfig_maximumRetryAttempts :: Lens' PutFunctionEventInvokeConfig (Maybe Natural)

The maximum number of times to retry when the function returns an error.

putFunctionEventInvokeConfig_functionName :: Lens' PutFunctionEventInvokeConfig Text

The name of the Lambda function, version, or alias.

Name formats

Function name - my-function (name-only), my-function:v1 (with alias).

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

You can append a version number or alias to any of the formats. The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

Destructuring the Response

data FunctionEventInvokeConfig

See: newFunctionEventInvokeConfig smart constructor.

Constructors

FunctionEventInvokeConfig'

Fields

maximumEventAgeInSeconds :: Maybe Natural

The maximum age of a request that Lambda sends to a function for processing.

functionArn :: Maybe Text

The Amazon Resource Name (ARN) of the function.

destinationConfig :: Maybe DestinationConfig

A destination for events after they have been sent to a function for processing.

Destinations

Function - The Amazon Resource Name (ARN) of a Lambda function.

Queue - The ARN of an SQS queue.

Topic - The ARN of an SNS topic.

Event Bus - The ARN of an Amazon EventBridge event bus.

maximumRetryAttempts :: Maybe Natural

The maximum number of times to retry when the function returns an error.

lastModified :: Maybe POSIX

The date and time that the configuration was last updated.

newFunctionEventInvokeConfig :: FunctionEventInvokeConfig

Create a value of FunctionEventInvokeConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maximumEventAgeInSeconds:FunctionEventInvokeConfig', functionEventInvokeConfig_maximumEventAgeInSeconds - The maximum age of a request that Lambda sends to a function for processing.

$sel:functionArn:FunctionEventInvokeConfig', functionEventInvokeConfig_functionArn - The Amazon Resource Name (ARN) of the function.

$sel:destinationConfig:FunctionEventInvokeConfig', functionEventInvokeConfig_destinationConfig - A destination for events after they have been sent to a function for processing.

Destinations

Function - The Amazon Resource Name (ARN) of a Lambda function.

Queue - The ARN of an SQS queue.

Topic - The ARN of an SNS topic.

Event Bus - The ARN of an Amazon EventBridge event bus.

$sel:maximumRetryAttempts:FunctionEventInvokeConfig', functionEventInvokeConfig_maximumRetryAttempts - The maximum number of times to retry when the function returns an error.

$sel:lastModified:FunctionEventInvokeConfig', functionEventInvokeConfig_lastModified - The date and time that the configuration was last updated.

Response Lenses

functionEventInvokeConfig_maximumEventAgeInSeconds :: Lens' FunctionEventInvokeConfig (Maybe Natural)

The maximum age of a request that Lambda sends to a function for processing.

functionEventInvokeConfig_functionArn :: Lens' FunctionEventInvokeConfig (Maybe Text)

The Amazon Resource Name (ARN) of the function.

functionEventInvokeConfig_destinationConfig :: Lens' FunctionEventInvokeConfig (Maybe DestinationConfig)

A destination for events after they have been sent to a function for processing.

Destinations

Function - The Amazon Resource Name (ARN) of a Lambda function.

Queue - The ARN of an SQS queue.

Topic - The ARN of an SNS topic.

Event Bus - The ARN of an Amazon EventBridge event bus.

functionEventInvokeConfig_maximumRetryAttempts :: Lens' FunctionEventInvokeConfig (Maybe Natural)

The maximum number of times to retry when the function returns an error.

functionEventInvokeConfig_lastModified :: Lens' FunctionEventInvokeConfig (Maybe UTCTime)

The date and time that the configuration was last updated.