Network.AWS.Lambda.PutProvisionedConcurrencyConfig

Description

Adds a provisioned concurrency configuration to a function's alias or version.

Creating a Request

data PutProvisionedConcurrencyConfig

See: newPutProvisionedConcurrencyConfig smart constructor.

Constructors

PutProvisionedConcurrencyConfig'

Fields

functionName :: Text

The name of the Lambda function.

Name formats

Function name - my-function.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

qualifier :: Text

The version number or alias name.

provisionedConcurrentExecutions :: Natural

The amount of provisioned concurrency to allocate for the version or alias.

newPutProvisionedConcurrencyConfig

Arguments

:: Text
$sel:functionName:PutProvisionedConcurrencyConfig'

-> Text
$sel:qualifier:PutProvisionedConcurrencyConfig'

-> Natural
$sel:provisionedConcurrentExecutions:PutProvisionedConcurrencyConfig'

-> PutProvisionedConcurrencyConfig

Create a value of PutProvisionedConcurrencyConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:functionName:PutProvisionedConcurrencyConfig', putProvisionedConcurrencyConfig_functionName - The name of the Lambda function.

Name formats

Function name - my-function.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

$sel:qualifier:PutProvisionedConcurrencyConfig', putProvisionedConcurrencyConfig_qualifier - The version number or alias name.

$sel:provisionedConcurrentExecutions:PutProvisionedConcurrencyConfig', putProvisionedConcurrencyConfig_provisionedConcurrentExecutions - The amount of provisioned concurrency to allocate for the version or alias.

Request Lenses

putProvisionedConcurrencyConfig_functionName :: Lens' PutProvisionedConcurrencyConfig Text

The name of the Lambda function.

Name formats

Function name - my-function.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:my-function.

Partial ARN - 123456789012:function:my-function.

The length constraint applies only to the full ARN. If you specify only the function name, it is limited to 64 characters in length.

putProvisionedConcurrencyConfig_qualifier :: Lens' PutProvisionedConcurrencyConfig Text

The version number or alias name.

putProvisionedConcurrencyConfig_provisionedConcurrentExecutions :: Lens' PutProvisionedConcurrencyConfig Natural

The amount of provisioned concurrency to allocate for the version or alias.

Destructuring the Response

data PutProvisionedConcurrencyConfigResponse

See: newPutProvisionedConcurrencyConfigResponse smart constructor.

Constructors

PutProvisionedConcurrencyConfigResponse'

Fields

status :: Maybe ProvisionedConcurrencyStatusEnum

The status of the allocation process.

availableProvisionedConcurrentExecutions :: Maybe Natural

The amount of provisioned concurrency available.

requestedProvisionedConcurrentExecutions :: Maybe Natural

The amount of provisioned concurrency requested.

allocatedProvisionedConcurrentExecutions :: Maybe Natural

The amount of provisioned concurrency allocated.

lastModified :: Maybe Text

The date and time that a user last updated the configuration, in ISO 8601 format.

statusReason :: Maybe Text

For failed allocations, the reason that provisioned concurrency could not be allocated.

httpStatus :: Int

The response's http status code.

newPutProvisionedConcurrencyConfigResponse

Arguments

:: Int
$sel:httpStatus:PutProvisionedConcurrencyConfigResponse'

-> PutProvisionedConcurrencyConfigResponse

Create a value of PutProvisionedConcurrencyConfigResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:PutProvisionedConcurrencyConfigResponse', putProvisionedConcurrencyConfigResponse_status - The status of the allocation process.

$sel:availableProvisionedConcurrentExecutions:PutProvisionedConcurrencyConfigResponse', putProvisionedConcurrencyConfigResponse_availableProvisionedConcurrentExecutions - The amount of provisioned concurrency available.

$sel:requestedProvisionedConcurrentExecutions:PutProvisionedConcurrencyConfigResponse', putProvisionedConcurrencyConfigResponse_requestedProvisionedConcurrentExecutions - The amount of provisioned concurrency requested.

$sel:allocatedProvisionedConcurrentExecutions:PutProvisionedConcurrencyConfigResponse', putProvisionedConcurrencyConfigResponse_allocatedProvisionedConcurrentExecutions - The amount of provisioned concurrency allocated.

$sel:lastModified:PutProvisionedConcurrencyConfigResponse', putProvisionedConcurrencyConfigResponse_lastModified - The date and time that a user last updated the configuration, in ISO 8601 format.

$sel:statusReason:PutProvisionedConcurrencyConfigResponse', putProvisionedConcurrencyConfigResponse_statusReason - For failed allocations, the reason that provisioned concurrency could not be allocated.

$sel:httpStatus:PutProvisionedConcurrencyConfigResponse', putProvisionedConcurrencyConfigResponse_httpStatus - The response's http status code.

Response Lenses

putProvisionedConcurrencyConfigResponse_status :: Lens' PutProvisionedConcurrencyConfigResponse (Maybe ProvisionedConcurrencyStatusEnum)

The status of the allocation process.

putProvisionedConcurrencyConfigResponse_availableProvisionedConcurrentExecutions :: Lens' PutProvisionedConcurrencyConfigResponse (Maybe Natural)

The amount of provisioned concurrency available.

putProvisionedConcurrencyConfigResponse_requestedProvisionedConcurrentExecutions :: Lens' PutProvisionedConcurrencyConfigResponse (Maybe Natural)

The amount of provisioned concurrency requested.

putProvisionedConcurrencyConfigResponse_allocatedProvisionedConcurrentExecutions :: Lens' PutProvisionedConcurrencyConfigResponse (Maybe Natural)

The amount of provisioned concurrency allocated.

putProvisionedConcurrencyConfigResponse_lastModified :: Lens' PutProvisionedConcurrencyConfigResponse (Maybe Text)

The date and time that a user last updated the configuration, in ISO 8601 format.

putProvisionedConcurrencyConfigResponse_statusReason :: Lens' PutProvisionedConcurrencyConfigResponse (Maybe Text)

For failed allocations, the reason that provisioned concurrency could not be allocated.

putProvisionedConcurrencyConfigResponse_httpStatus :: Lens' PutProvisionedConcurrencyConfigResponse Int

The response's http status code.