Network.AWS.Lambda.Types.AccountLimit

Description

Documentation

data AccountLimit

Limits that are related to concurrency and storage. All file and storage sizes are in bytes.

See: newAccountLimit smart constructor.

Constructors

AccountLimit'

Fields

codeSizeUnzipped :: Maybe Integer

The maximum size of a function's deployment package and layers when they're extracted.

concurrentExecutions :: Maybe Int

The maximum number of simultaneous function executions.

unreservedConcurrentExecutions :: Maybe Natural

The maximum number of simultaneous function executions, minus the capacity that's reserved for individual functions with PutFunctionConcurrency.

totalCodeSize :: Maybe Integer

The amount of storage space that you can use for all deployment packages and layer archives.

codeSizeZipped :: Maybe Integer

The maximum size of a deployment package when it's uploaded directly to Lambda. Use Amazon S3 for larger files.

newAccountLimit :: AccountLimit

Create a value of AccountLimit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:codeSizeUnzipped:AccountLimit', accountLimit_codeSizeUnzipped - The maximum size of a function's deployment package and layers when they're extracted.

$sel:concurrentExecutions:AccountLimit', accountLimit_concurrentExecutions - The maximum number of simultaneous function executions.

$sel:unreservedConcurrentExecutions:AccountLimit', accountLimit_unreservedConcurrentExecutions - The maximum number of simultaneous function executions, minus the capacity that's reserved for individual functions with PutFunctionConcurrency.

$sel:totalCodeSize:AccountLimit', accountLimit_totalCodeSize - The amount of storage space that you can use for all deployment packages and layer archives.

$sel:codeSizeZipped:AccountLimit', accountLimit_codeSizeZipped - The maximum size of a deployment package when it's uploaded directly to Lambda. Use Amazon S3 for larger files.

accountLimit_codeSizeUnzipped :: Lens' AccountLimit (Maybe Integer)

The maximum size of a function's deployment package and layers when they're extracted.

accountLimit_concurrentExecutions :: Lens' AccountLimit (Maybe Int)

The maximum number of simultaneous function executions.

accountLimit_unreservedConcurrentExecutions :: Lens' AccountLimit (Maybe Natural)

The maximum number of simultaneous function executions, minus the capacity that's reserved for individual functions with PutFunctionConcurrency.

accountLimit_totalCodeSize :: Lens' AccountLimit (Maybe Integer)

The amount of storage space that you can use for all deployment packages and layer archives.

accountLimit_codeSizeZipped :: Lens' AccountLimit (Maybe Integer)

The maximum size of a deployment package when it's uploaded directly to Lambda. Use Amazon S3 for larger files.