Network.AWS.Lambda.Types.ProvisionedConcurrencyConfigListItem

Description

Documentation

data ProvisionedConcurrencyConfigListItem

Details about the provisioned concurrency configuration for a function alias or version.

See: newProvisionedConcurrencyConfigListItem smart constructor.

Constructors

ProvisionedConcurrencyConfigListItem'

Fields

status :: Maybe ProvisionedConcurrencyStatusEnum

The status of the allocation process.

availableProvisionedConcurrentExecutions :: Maybe Natural

The amount of provisioned concurrency available.

requestedProvisionedConcurrentExecutions :: Maybe Natural

The amount of provisioned concurrency requested.

functionArn :: Maybe Text

The Amazon Resource Name (ARN) of the alias or version.

allocatedProvisionedConcurrentExecutions :: Maybe Natural

The amount of provisioned concurrency allocated.

lastModified :: Maybe Text

The date and time that a user last updated the configuration, in ISO 8601 format.

statusReason :: Maybe Text

For failed allocations, the reason that provisioned concurrency could not be allocated.

newProvisionedConcurrencyConfigListItem :: ProvisionedConcurrencyConfigListItem

Create a value of ProvisionedConcurrencyConfigListItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ProvisionedConcurrencyConfigListItem', provisionedConcurrencyConfigListItem_status - The status of the allocation process.

$sel:availableProvisionedConcurrentExecutions:ProvisionedConcurrencyConfigListItem', provisionedConcurrencyConfigListItem_availableProvisionedConcurrentExecutions - The amount of provisioned concurrency available.

$sel:requestedProvisionedConcurrentExecutions:ProvisionedConcurrencyConfigListItem', provisionedConcurrencyConfigListItem_requestedProvisionedConcurrentExecutions - The amount of provisioned concurrency requested.

$sel:functionArn:ProvisionedConcurrencyConfigListItem', provisionedConcurrencyConfigListItem_functionArn - The Amazon Resource Name (ARN) of the alias or version.

$sel:allocatedProvisionedConcurrentExecutions:ProvisionedConcurrencyConfigListItem', provisionedConcurrencyConfigListItem_allocatedProvisionedConcurrentExecutions - The amount of provisioned concurrency allocated.

$sel:lastModified:ProvisionedConcurrencyConfigListItem', provisionedConcurrencyConfigListItem_lastModified - The date and time that a user last updated the configuration, in ISO 8601 format.

$sel:statusReason:ProvisionedConcurrencyConfigListItem', provisionedConcurrencyConfigListItem_statusReason - For failed allocations, the reason that provisioned concurrency could not be allocated.

provisionedConcurrencyConfigListItem_status :: Lens' ProvisionedConcurrencyConfigListItem (Maybe ProvisionedConcurrencyStatusEnum)

The status of the allocation process.

provisionedConcurrencyConfigListItem_availableProvisionedConcurrentExecutions :: Lens' ProvisionedConcurrencyConfigListItem (Maybe Natural)

The amount of provisioned concurrency available.

provisionedConcurrencyConfigListItem_requestedProvisionedConcurrentExecutions :: Lens' ProvisionedConcurrencyConfigListItem (Maybe Natural)

The amount of provisioned concurrency requested.

provisionedConcurrencyConfigListItem_functionArn :: Lens' ProvisionedConcurrencyConfigListItem (Maybe Text)

The Amazon Resource Name (ARN) of the alias or version.

provisionedConcurrencyConfigListItem_allocatedProvisionedConcurrentExecutions :: Lens' ProvisionedConcurrencyConfigListItem (Maybe Natural)

The amount of provisioned concurrency allocated.

provisionedConcurrencyConfigListItem_lastModified :: Lens' ProvisionedConcurrencyConfigListItem (Maybe Text)

The date and time that a user last updated the configuration, in ISO 8601 format.

provisionedConcurrencyConfigListItem_statusReason :: Lens' ProvisionedConcurrencyConfigListItem (Maybe Text)

For failed allocations, the reason that provisioned concurrency could not be allocated.