Network.AWS.Lambda.UpdateEventSourceMapping

Description

Updates an event source mapping. You can change the function that Lambda invokes, or pause invocation and resume later from the same location.

The following error handling options are only available for stream sources (DynamoDB and Kinesis):

BisectBatchOnFunctionError - If the function returns an error, split the batch in two and retry.

DestinationConfig - Send discarded records to an Amazon SQS queue or Amazon SNS topic.

MaximumRecordAgeInSeconds - Discard records older than the specified age. The default value is infinite (-1). When set to infinite (-1), failed records are retried until the record expires

MaximumRetryAttempts - Discard records after the specified number of retries. The default value is infinite (-1). When set to infinite (-1), failed records are retried until the record expires.

ParallelizationFactor - Process multiple batches from each shard concurrently.

Creating a Request

data UpdateEventSourceMapping

See: newUpdateEventSourceMapping smart constructor.

Constructors

UpdateEventSourceMapping'

Fields

enabled :: Maybe Bool

If true, the event source mapping is active. Set to false to pause polling and invocation.

functionName :: Maybe Text

The name of the Lambda function.

Name formats

Function name - MyFunction.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction.

Version or Alias ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction:PROD.

Partial ARN - 123456789012:function:MyFunction.

The length constraint applies only to the full ARN. If you specify only the function name, it's limited to 64 characters in length.

maximumRecordAgeInSeconds :: Maybe Int

(Streams only) Discard records older than the specified age. The default value is infinite (-1).

functionResponseTypes :: Maybe [FunctionResponseType]

(Streams only) A list of current response type enums applied to the event source mapping.

tumblingWindowInSeconds :: Maybe Natural

(Streams only) The duration in seconds of a processing window. The range is between 1 second up to 900 seconds.

maximumBatchingWindowInSeconds :: Maybe Natural

(Streams and SQS standard queues) The maximum amount of time to gather records before invoking the function, in seconds.

batchSize :: Maybe Natural

The maximum number of items to retrieve in a single batch.

Amazon Kinesis - Default 100. Max 10,000.

Amazon DynamoDB Streams - Default 100. Max 1,000.

Amazon Simple Queue Service - Default 10. For standard queues the max is 10,000. For FIFO queues the max is 10.

Amazon Managed Streaming for Apache Kafka - Default 100. Max 10,000.

Self-Managed Apache Kafka - Default 100. Max 10,000.

destinationConfig :: Maybe DestinationConfig

(Streams only) An Amazon SQS queue or Amazon SNS topic destination for discarded records.

maximumRetryAttempts :: Maybe Int

(Streams only) Discard records after the specified number of retries. The default value is infinite (-1). When set to infinite (-1), failed records will be retried until the record expires.

parallelizationFactor :: Maybe Natural

(Streams only) The number of batches to process from each shard concurrently.

bisectBatchOnFunctionError :: Maybe Bool

(Streams only) If the function returns an error, split the batch in two and retry.

sourceAccessConfigurations :: Maybe [SourceAccessConfiguration]

An array of authentication protocols or VPC components required to secure your event source.

uuid :: Text

The identifier of the event source mapping.

newUpdateEventSourceMapping

Arguments

:: Text
$sel:uuid:UpdateEventSourceMapping'

-> UpdateEventSourceMapping

Create a value of UpdateEventSourceMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:UpdateEventSourceMapping', updateEventSourceMapping_enabled - If true, the event source mapping is active. Set to false to pause polling and invocation.

$sel:functionName:UpdateEventSourceMapping', updateEventSourceMapping_functionName - The name of the Lambda function.

Name formats

Function name - MyFunction.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction.

Version or Alias ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction:PROD.

Partial ARN - 123456789012:function:MyFunction.

The length constraint applies only to the full ARN. If you specify only the function name, it's limited to 64 characters in length.

$sel:maximumRecordAgeInSeconds:UpdateEventSourceMapping', updateEventSourceMapping_maximumRecordAgeInSeconds - (Streams only) Discard records older than the specified age. The default value is infinite (-1).

$sel:functionResponseTypes:UpdateEventSourceMapping', updateEventSourceMapping_functionResponseTypes - (Streams only) A list of current response type enums applied to the event source mapping.

$sel:tumblingWindowInSeconds:UpdateEventSourceMapping', updateEventSourceMapping_tumblingWindowInSeconds - (Streams only) The duration in seconds of a processing window. The range is between 1 second up to 900 seconds.

$sel:maximumBatchingWindowInSeconds:UpdateEventSourceMapping', updateEventSourceMapping_maximumBatchingWindowInSeconds - (Streams and SQS standard queues) The maximum amount of time to gather records before invoking the function, in seconds.

$sel:batchSize:UpdateEventSourceMapping', updateEventSourceMapping_batchSize - The maximum number of items to retrieve in a single batch.

Amazon Kinesis - Default 100. Max 10,000.

Amazon DynamoDB Streams - Default 100. Max 1,000.

Amazon Simple Queue Service - Default 10. For standard queues the max is 10,000. For FIFO queues the max is 10.

Amazon Managed Streaming for Apache Kafka - Default 100. Max 10,000.

Self-Managed Apache Kafka - Default 100. Max 10,000.

$sel:destinationConfig:UpdateEventSourceMapping', updateEventSourceMapping_destinationConfig - (Streams only) An Amazon SQS queue or Amazon SNS topic destination for discarded records.

$sel:maximumRetryAttempts:UpdateEventSourceMapping', updateEventSourceMapping_maximumRetryAttempts - (Streams only) Discard records after the specified number of retries. The default value is infinite (-1). When set to infinite (-1), failed records will be retried until the record expires.

$sel:parallelizationFactor:UpdateEventSourceMapping', updateEventSourceMapping_parallelizationFactor - (Streams only) The number of batches to process from each shard concurrently.

$sel:bisectBatchOnFunctionError:UpdateEventSourceMapping', updateEventSourceMapping_bisectBatchOnFunctionError - (Streams only) If the function returns an error, split the batch in two and retry.

$sel:sourceAccessConfigurations:UpdateEventSourceMapping', updateEventSourceMapping_sourceAccessConfigurations - An array of authentication protocols or VPC components required to secure your event source.

$sel:uuid:UpdateEventSourceMapping', updateEventSourceMapping_uuid - The identifier of the event source mapping.

Request Lenses

updateEventSourceMapping_enabled :: Lens' UpdateEventSourceMapping (Maybe Bool)

If true, the event source mapping is active. Set to false to pause polling and invocation.

updateEventSourceMapping_functionName :: Lens' UpdateEventSourceMapping (Maybe Text)

The name of the Lambda function.

Name formats

Function name - MyFunction.

Function ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction.

Version or Alias ARN - arn:aws:lambda:us-west-2:123456789012:function:MyFunction:PROD.

Partial ARN - 123456789012:function:MyFunction.

The length constraint applies only to the full ARN. If you specify only the function name, it's limited to 64 characters in length.

updateEventSourceMapping_maximumRecordAgeInSeconds :: Lens' UpdateEventSourceMapping (Maybe Int)

(Streams only) Discard records older than the specified age. The default value is infinite (-1).

updateEventSourceMapping_functionResponseTypes :: Lens' UpdateEventSourceMapping (Maybe [FunctionResponseType])

(Streams only) A list of current response type enums applied to the event source mapping.

updateEventSourceMapping_tumblingWindowInSeconds :: Lens' UpdateEventSourceMapping (Maybe Natural)

(Streams only) The duration in seconds of a processing window. The range is between 1 second up to 900 seconds.

updateEventSourceMapping_maximumBatchingWindowInSeconds :: Lens' UpdateEventSourceMapping (Maybe Natural)

(Streams and SQS standard queues) The maximum amount of time to gather records before invoking the function, in seconds.

updateEventSourceMapping_batchSize :: Lens' UpdateEventSourceMapping (Maybe Natural)

The maximum number of items to retrieve in a single batch.

Amazon Kinesis - Default 100. Max 10,000.

Amazon DynamoDB Streams - Default 100. Max 1,000.

Amazon Simple Queue Service - Default 10. For standard queues the max is 10,000. For FIFO queues the max is 10.

Amazon Managed Streaming for Apache Kafka - Default 100. Max 10,000.

Self-Managed Apache Kafka - Default 100. Max 10,000.

updateEventSourceMapping_destinationConfig :: Lens' UpdateEventSourceMapping (Maybe DestinationConfig)

(Streams only) An Amazon SQS queue or Amazon SNS topic destination for discarded records.

updateEventSourceMapping_maximumRetryAttempts :: Lens' UpdateEventSourceMapping (Maybe Int)

(Streams only) Discard records after the specified number of retries. The default value is infinite (-1). When set to infinite (-1), failed records will be retried until the record expires.

updateEventSourceMapping_parallelizationFactor :: Lens' UpdateEventSourceMapping (Maybe Natural)

(Streams only) The number of batches to process from each shard concurrently.

updateEventSourceMapping_bisectBatchOnFunctionError :: Lens' UpdateEventSourceMapping (Maybe Bool)

(Streams only) If the function returns an error, split the batch in two and retry.

updateEventSourceMapping_sourceAccessConfigurations :: Lens' UpdateEventSourceMapping (Maybe [SourceAccessConfiguration])

An array of authentication protocols or VPC components required to secure your event source.

updateEventSourceMapping_uuid :: Lens' UpdateEventSourceMapping Text

The identifier of the event source mapping.

Destructuring the Response

data EventSourceMappingConfiguration

A mapping between an Amazon Web Services resource and a Lambda function. For details, see CreateEventSourceMapping.

See: newEventSourceMappingConfiguration smart constructor.

Constructors

EventSourceMappingConfiguration'

Fields

eventSourceArn :: Maybe Text

The Amazon Resource Name (ARN) of the event source.

stateTransitionReason :: Maybe Text

Indicates whether a user or Lambda made the last change to the event source mapping.

lastProcessingResult :: Maybe Text

The result of the last Lambda invocation of your function.

queues :: Maybe (NonEmpty Text)

(Amazon MQ) The name of the Amazon MQ broker destination queue to consume.

maximumRecordAgeInSeconds :: Maybe Int

(Streams only) Discard records older than the specified age. The default value is -1, which sets the maximum age to infinite. When the value is set to infinite, Lambda never discards old records.

topics :: Maybe (NonEmpty Text)

The name of the Kafka topic.

functionResponseTypes :: Maybe [FunctionResponseType]

(Streams only) A list of current response type enums applied to the event source mapping.

tumblingWindowInSeconds :: Maybe Natural

(Streams only) The duration in seconds of a processing window. The range is 1–900 seconds.

startingPositionTimestamp :: Maybe POSIX

With StartingPosition set to AT_TIMESTAMP, the time from which to start reading.

state :: Maybe Text

The state of the event source mapping. It can be one of the following: Creating, Enabling, Enabled, Disabling, Disabled, Updating, or Deleting.

functionArn :: Maybe Text

The ARN of the Lambda function.

maximumBatchingWindowInSeconds :: Maybe Natural

(Streams and Amazon SQS standard queues) The maximum amount of time to gather records before invoking the function, in seconds. The default value is zero.

batchSize :: Maybe Natural

The maximum number of items to retrieve in a single batch.

startingPosition :: Maybe EventSourcePosition

The position in a stream from which to start reading. Required for Amazon Kinesis, Amazon DynamoDB, and Amazon MSK stream sources. AT_TIMESTAMP is supported only for Amazon Kinesis streams.

destinationConfig :: Maybe DestinationConfig

(Streams only) An Amazon SQS queue or Amazon SNS topic destination for discarded records.

maximumRetryAttempts :: Maybe Int

(Streams only) Discard records after the specified number of retries. The default value is -1, which sets the maximum number of retries to infinite. When MaximumRetryAttempts is infinite, Lambda retries failed records until the record expires in the event source.

lastModified :: Maybe POSIX

The date that the event source mapping was last updated or that its state changed.

selfManagedEventSource :: Maybe SelfManagedEventSource

The self-managed Apache Kafka cluster for your event source.

parallelizationFactor :: Maybe Natural

(Streams only) The number of batches to process concurrently from each shard. The default value is 1.

bisectBatchOnFunctionError :: Maybe Bool

(Streams only) If the function returns an error, split the batch in two and retry. The default value is false.

uuid :: Maybe Text

The identifier of the event source mapping.

sourceAccessConfigurations :: Maybe [SourceAccessConfiguration]

An array of the authentication protocol, VPC components, or virtual host to secure and define your event source.

newEventSourceMappingConfiguration :: EventSourceMappingConfiguration

Create a value of EventSourceMappingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventSourceArn:EventSourceMappingConfiguration', eventSourceMappingConfiguration_eventSourceArn - The Amazon Resource Name (ARN) of the event source.

$sel:stateTransitionReason:EventSourceMappingConfiguration', eventSourceMappingConfiguration_stateTransitionReason - Indicates whether a user or Lambda made the last change to the event source mapping.

$sel:lastProcessingResult:EventSourceMappingConfiguration', eventSourceMappingConfiguration_lastProcessingResult - The result of the last Lambda invocation of your function.

$sel:queues:EventSourceMappingConfiguration', eventSourceMappingConfiguration_queues - (Amazon MQ) The name of the Amazon MQ broker destination queue to consume.

$sel:maximumRecordAgeInSeconds:EventSourceMappingConfiguration', eventSourceMappingConfiguration_maximumRecordAgeInSeconds - (Streams only) Discard records older than the specified age. The default value is -1, which sets the maximum age to infinite. When the value is set to infinite, Lambda never discards old records.

$sel:topics:EventSourceMappingConfiguration', eventSourceMappingConfiguration_topics - The name of the Kafka topic.

$sel:functionResponseTypes:EventSourceMappingConfiguration', eventSourceMappingConfiguration_functionResponseTypes - (Streams only) A list of current response type enums applied to the event source mapping.

$sel:tumblingWindowInSeconds:EventSourceMappingConfiguration', eventSourceMappingConfiguration_tumblingWindowInSeconds - (Streams only) The duration in seconds of a processing window. The range is 1–900 seconds.

$sel:startingPositionTimestamp:EventSourceMappingConfiguration', eventSourceMappingConfiguration_startingPositionTimestamp - With StartingPosition set to AT_TIMESTAMP, the time from which to start reading.

$sel:state:EventSourceMappingConfiguration', eventSourceMappingConfiguration_state - The state of the event source mapping. It can be one of the following: Creating, Enabling, Enabled, Disabling, Disabled, Updating, or Deleting.

$sel:functionArn:EventSourceMappingConfiguration', eventSourceMappingConfiguration_functionArn - The ARN of the Lambda function.

$sel:maximumBatchingWindowInSeconds:EventSourceMappingConfiguration', eventSourceMappingConfiguration_maximumBatchingWindowInSeconds - (Streams and Amazon SQS standard queues) The maximum amount of time to gather records before invoking the function, in seconds. The default value is zero.

$sel:batchSize:EventSourceMappingConfiguration', eventSourceMappingConfiguration_batchSize - The maximum number of items to retrieve in a single batch.

$sel:startingPosition:EventSourceMappingConfiguration', eventSourceMappingConfiguration_startingPosition - The position in a stream from which to start reading. Required for Amazon Kinesis, Amazon DynamoDB, and Amazon MSK stream sources. AT_TIMESTAMP is supported only for Amazon Kinesis streams.

$sel:destinationConfig:EventSourceMappingConfiguration', eventSourceMappingConfiguration_destinationConfig - (Streams only) An Amazon SQS queue or Amazon SNS topic destination for discarded records.

$sel:maximumRetryAttempts:EventSourceMappingConfiguration', eventSourceMappingConfiguration_maximumRetryAttempts - (Streams only) Discard records after the specified number of retries. The default value is -1, which sets the maximum number of retries to infinite. When MaximumRetryAttempts is infinite, Lambda retries failed records until the record expires in the event source.

$sel:lastModified:EventSourceMappingConfiguration', eventSourceMappingConfiguration_lastModified - The date that the event source mapping was last updated or that its state changed.

$sel:selfManagedEventSource:EventSourceMappingConfiguration', eventSourceMappingConfiguration_selfManagedEventSource - The self-managed Apache Kafka cluster for your event source.

$sel:parallelizationFactor:EventSourceMappingConfiguration', eventSourceMappingConfiguration_parallelizationFactor - (Streams only) The number of batches to process concurrently from each shard. The default value is 1.

$sel:bisectBatchOnFunctionError:EventSourceMappingConfiguration', eventSourceMappingConfiguration_bisectBatchOnFunctionError - (Streams only) If the function returns an error, split the batch in two and retry. The default value is false.

$sel:uuid:EventSourceMappingConfiguration', eventSourceMappingConfiguration_uuid - The identifier of the event source mapping.

$sel:sourceAccessConfigurations:EventSourceMappingConfiguration', eventSourceMappingConfiguration_sourceAccessConfigurations - An array of the authentication protocol, VPC components, or virtual host to secure and define your event source.

Response Lenses

eventSourceMappingConfiguration_eventSourceArn :: Lens' EventSourceMappingConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the event source.

eventSourceMappingConfiguration_stateTransitionReason :: Lens' EventSourceMappingConfiguration (Maybe Text)

Indicates whether a user or Lambda made the last change to the event source mapping.

eventSourceMappingConfiguration_lastProcessingResult :: Lens' EventSourceMappingConfiguration (Maybe Text)

The result of the last Lambda invocation of your function.

eventSourceMappingConfiguration_queues :: Lens' EventSourceMappingConfiguration (Maybe (NonEmpty Text))

(Amazon MQ) The name of the Amazon MQ broker destination queue to consume.

eventSourceMappingConfiguration_maximumRecordAgeInSeconds :: Lens' EventSourceMappingConfiguration (Maybe Int)

(Streams only) Discard records older than the specified age. The default value is -1, which sets the maximum age to infinite. When the value is set to infinite, Lambda never discards old records.

eventSourceMappingConfiguration_topics :: Lens' EventSourceMappingConfiguration (Maybe (NonEmpty Text))

The name of the Kafka topic.

eventSourceMappingConfiguration_functionResponseTypes :: Lens' EventSourceMappingConfiguration (Maybe [FunctionResponseType])

(Streams only) A list of current response type enums applied to the event source mapping.

eventSourceMappingConfiguration_tumblingWindowInSeconds :: Lens' EventSourceMappingConfiguration (Maybe Natural)

(Streams only) The duration in seconds of a processing window. The range is 1–900 seconds.

eventSourceMappingConfiguration_startingPositionTimestamp :: Lens' EventSourceMappingConfiguration (Maybe UTCTime)

With StartingPosition set to AT_TIMESTAMP, the time from which to start reading.

eventSourceMappingConfiguration_state :: Lens' EventSourceMappingConfiguration (Maybe Text)

The state of the event source mapping. It can be one of the following: Creating, Enabling, Enabled, Disabling, Disabled, Updating, or Deleting.

eventSourceMappingConfiguration_functionArn :: Lens' EventSourceMappingConfiguration (Maybe Text)

The ARN of the Lambda function.

eventSourceMappingConfiguration_maximumBatchingWindowInSeconds :: Lens' EventSourceMappingConfiguration (Maybe Natural)

(Streams and Amazon SQS standard queues) The maximum amount of time to gather records before invoking the function, in seconds. The default value is zero.

eventSourceMappingConfiguration_batchSize :: Lens' EventSourceMappingConfiguration (Maybe Natural)

The maximum number of items to retrieve in a single batch.

eventSourceMappingConfiguration_startingPosition :: Lens' EventSourceMappingConfiguration (Maybe EventSourcePosition)

The position in a stream from which to start reading. Required for Amazon Kinesis, Amazon DynamoDB, and Amazon MSK stream sources. AT_TIMESTAMP is supported only for Amazon Kinesis streams.

eventSourceMappingConfiguration_destinationConfig :: Lens' EventSourceMappingConfiguration (Maybe DestinationConfig)

(Streams only) An Amazon SQS queue or Amazon SNS topic destination for discarded records.

eventSourceMappingConfiguration_maximumRetryAttempts :: Lens' EventSourceMappingConfiguration (Maybe Int)

(Streams only) Discard records after the specified number of retries. The default value is -1, which sets the maximum number of retries to infinite. When MaximumRetryAttempts is infinite, Lambda retries failed records until the record expires in the event source.

eventSourceMappingConfiguration_lastModified :: Lens' EventSourceMappingConfiguration (Maybe UTCTime)

The date that the event source mapping was last updated or that its state changed.

eventSourceMappingConfiguration_selfManagedEventSource :: Lens' EventSourceMappingConfiguration (Maybe SelfManagedEventSource)

The self-managed Apache Kafka cluster for your event source.

eventSourceMappingConfiguration_parallelizationFactor :: Lens' EventSourceMappingConfiguration (Maybe Natural)

(Streams only) The number of batches to process concurrently from each shard. The default value is 1.

eventSourceMappingConfiguration_bisectBatchOnFunctionError :: Lens' EventSourceMappingConfiguration (Maybe Bool)

(Streams only) If the function returns an error, split the batch in two and retry. The default value is false.

eventSourceMappingConfiguration_uuid :: Lens' EventSourceMappingConfiguration (Maybe Text)

The identifier of the event source mapping.

eventSourceMappingConfiguration_sourceAccessConfigurations :: Lens' EventSourceMappingConfiguration (Maybe [SourceAccessConfiguration])

An array of the authentication protocol, VPC components, or virtual host to secure and define your event source.