Network.AWS.LexModels.CreateBotVersion

Description

Creates a new version of the bot based on the $LATEST version. If the $LATEST version of this resource hasn't changed since you created the last version, Amazon Lex doesn't create a new version. It returns the last created version.

You can update only the $LATEST version of the bot. You can't update the numbered versions that you create with the CreateBotVersion operation.

When you create the first version of a bot, Amazon Lex sets the version to 1. Subsequent versions increment by 1. For more information, see versioning-intro.

This operation requires permission for the lex:CreateBotVersion action.

Creating a Request

data CreateBotVersion

See: newCreateBotVersion smart constructor.

Constructors

CreateBotVersion'

Fields

checksum :: Maybe Text

Identifies a specific revision of the $LATEST version of the bot. If you specify a checksum and the $LATEST version of the bot has a different checksum, a PreconditionFailedException exception is returned and Amazon Lex doesn't publish a new version. If you don't specify a checksum, Amazon Lex publishes the $LATEST version.

name :: Text

The name of the bot that you want to create a new version of. The name is case sensitive.

newCreateBotVersion

Arguments

:: Text
$sel:name:CreateBotVersion'

-> CreateBotVersion

Create a value of CreateBotVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checksum:CreateBotVersion', createBotVersion_checksum - Identifies a specific revision of the $LATEST version of the bot. If you specify a checksum and the $LATEST version of the bot has a different checksum, a PreconditionFailedException exception is returned and Amazon Lex doesn't publish a new version. If you don't specify a checksum, Amazon Lex publishes the $LATEST version.

$sel:name:CreateBotVersion', createBotVersion_name - The name of the bot that you want to create a new version of. The name is case sensitive.

Request Lenses

createBotVersion_checksum :: Lens' CreateBotVersion (Maybe Text)

Identifies a specific revision of the $LATEST version of the bot. If you specify a checksum and the $LATEST version of the bot has a different checksum, a PreconditionFailedException exception is returned and Amazon Lex doesn't publish a new version. If you don't specify a checksum, Amazon Lex publishes the $LATEST version.

createBotVersion_name :: Lens' CreateBotVersion Text

The name of the bot that you want to create a new version of. The name is case sensitive.

Destructuring the Response

data CreateBotVersionResponse

See: newCreateBotVersionResponse smart constructor.

Constructors

CreateBotVersionResponse'

Fields

failureReason :: Maybe Text

If status is FAILED, Amazon Lex provides the reason that it failed to build the bot.

status :: Maybe LexStatus

When you send a request to create or update a bot, Amazon Lex sets the status response element to BUILDING. After Amazon Lex builds the bot, it sets status to READY. If Amazon Lex can't build the bot, it sets status to FAILED. Amazon Lex returns the reason for the failure in the failureReason response element.

abortStatement :: Maybe Statement

The message that Amazon Lex uses to cancel a conversation. For more information, see PutBot.

intents :: Maybe [Intent]

An array of Intent objects. For more information, see PutBot.

checksum :: Maybe Text

Checksum identifying the version of the bot that was created.

enableModelImprovements :: Maybe Bool

Indicates whether the bot uses accuracy improvements. true indicates that the bot is using the improvements, otherwise, false.

detectSentiment :: Maybe Bool

Indicates whether utterances entered by the user should be sent to Amazon Comprehend for sentiment analysis.

locale :: Maybe Locale

Specifies the target locale for the bot.

createdDate :: Maybe POSIX

The date when the bot version was created.

name :: Maybe Text

The name of the bot.

version :: Maybe Text

The version of the bot.

idleSessionTTLInSeconds :: Maybe Natural

The maximum time in seconds that Amazon Lex retains the data gathered in a conversation. For more information, see PutBot.

clarificationPrompt :: Maybe Prompt

The message that Amazon Lex uses when it doesn't understand the user's request. For more information, see PutBot.

voiceId :: Maybe Text

The Amazon Polly voice ID that Amazon Lex uses for voice interactions with the user.

lastUpdatedDate :: Maybe POSIX

The date when the $LATEST version of this bot was updated.

childDirected :: Maybe Bool

For each Amazon Lex bot created with the Amazon Lex Model Building Service, you must specify whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to the Children's Online Privacy Protection Act (COPPA) by specifying true or false in the childDirected field. By specifying true in the childDirected field, you confirm that your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. By specifying false in the childDirected field, you confirm that your use of Amazon Lex is not related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. You may not specify a default value for the childDirected field that does not accurately reflect whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA.

If your use of Amazon Lex relates to a website, program, or other application that is directed in whole or in part, to children under age 13, you must obtain any required verifiable parental consent under COPPA. For information regarding the use of Amazon Lex in connection with websites, programs, or other applications that are directed or targeted, in whole or in part, to children under age 13, see the Amazon Lex FAQ.

description :: Maybe Text

A description of the bot.

httpStatus :: Int

The response's http status code.

newCreateBotVersionResponse

Arguments

:: Int
$sel:httpStatus:CreateBotVersionResponse'

-> CreateBotVersionResponse

Create a value of CreateBotVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:CreateBotVersionResponse', createBotVersionResponse_failureReason - If status is FAILED, Amazon Lex provides the reason that it failed to build the bot.

$sel:status:CreateBotVersionResponse', createBotVersionResponse_status - When you send a request to create or update a bot, Amazon Lex sets the status response element to BUILDING. After Amazon Lex builds the bot, it sets status to READY. If Amazon Lex can't build the bot, it sets status to FAILED. Amazon Lex returns the reason for the failure in the failureReason response element.

$sel:abortStatement:CreateBotVersionResponse', createBotVersionResponse_abortStatement - The message that Amazon Lex uses to cancel a conversation. For more information, see PutBot.

$sel:intents:CreateBotVersionResponse', createBotVersionResponse_intents - An array of Intent objects. For more information, see PutBot.

$sel:checksum:CreateBotVersion', createBotVersionResponse_checksum - Checksum identifying the version of the bot that was created.

$sel:enableModelImprovements:CreateBotVersionResponse', createBotVersionResponse_enableModelImprovements - Indicates whether the bot uses accuracy improvements. true indicates that the bot is using the improvements, otherwise, false.

$sel:detectSentiment:CreateBotVersionResponse', createBotVersionResponse_detectSentiment - Indicates whether utterances entered by the user should be sent to Amazon Comprehend for sentiment analysis.

$sel:locale:CreateBotVersionResponse', createBotVersionResponse_locale - Specifies the target locale for the bot.

$sel:createdDate:CreateBotVersionResponse', createBotVersionResponse_createdDate - The date when the bot version was created.

$sel:name:CreateBotVersion', createBotVersionResponse_name - The name of the bot.

$sel:version:CreateBotVersionResponse', createBotVersionResponse_version - The version of the bot.

$sel:idleSessionTTLInSeconds:CreateBotVersionResponse', createBotVersionResponse_idleSessionTTLInSeconds - The maximum time in seconds that Amazon Lex retains the data gathered in a conversation. For more information, see PutBot.

$sel:clarificationPrompt:CreateBotVersionResponse', createBotVersionResponse_clarificationPrompt - The message that Amazon Lex uses when it doesn't understand the user's request. For more information, see PutBot.

$sel:voiceId:CreateBotVersionResponse', createBotVersionResponse_voiceId - The Amazon Polly voice ID that Amazon Lex uses for voice interactions with the user.

$sel:lastUpdatedDate:CreateBotVersionResponse', createBotVersionResponse_lastUpdatedDate - The date when the $LATEST version of this bot was updated.

$sel:childDirected:CreateBotVersionResponse', createBotVersionResponse_childDirected - For each Amazon Lex bot created with the Amazon Lex Model Building Service, you must specify whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to the Children's Online Privacy Protection Act (COPPA) by specifying true or false in the childDirected field. By specifying true in the childDirected field, you confirm that your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. By specifying false in the childDirected field, you confirm that your use of Amazon Lex is not related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. You may not specify a default value for the childDirected field that does not accurately reflect whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA.

If your use of Amazon Lex relates to a website, program, or other application that is directed in whole or in part, to children under age 13, you must obtain any required verifiable parental consent under COPPA. For information regarding the use of Amazon Lex in connection with websites, programs, or other applications that are directed or targeted, in whole or in part, to children under age 13, see the Amazon Lex FAQ.

$sel:description:CreateBotVersionResponse', createBotVersionResponse_description - A description of the bot.

$sel:httpStatus:CreateBotVersionResponse', createBotVersionResponse_httpStatus - The response's http status code.

Response Lenses

createBotVersionResponse_failureReason :: Lens' CreateBotVersionResponse (Maybe Text)

If status is FAILED, Amazon Lex provides the reason that it failed to build the bot.

createBotVersionResponse_status :: Lens' CreateBotVersionResponse (Maybe LexStatus)

When you send a request to create or update a bot, Amazon Lex sets the status response element to BUILDING. After Amazon Lex builds the bot, it sets status to READY. If Amazon Lex can't build the bot, it sets status to FAILED. Amazon Lex returns the reason for the failure in the failureReason response element.

createBotVersionResponse_abortStatement :: Lens' CreateBotVersionResponse (Maybe Statement)

The message that Amazon Lex uses to cancel a conversation. For more information, see PutBot.

createBotVersionResponse_intents :: Lens' CreateBotVersionResponse (Maybe [Intent])

An array of Intent objects. For more information, see PutBot.

createBotVersionResponse_checksum :: Lens' CreateBotVersionResponse (Maybe Text)

Checksum identifying the version of the bot that was created.

createBotVersionResponse_enableModelImprovements :: Lens' CreateBotVersionResponse (Maybe Bool)

Indicates whether the bot uses accuracy improvements. true indicates that the bot is using the improvements, otherwise, false.

createBotVersionResponse_detectSentiment :: Lens' CreateBotVersionResponse (Maybe Bool)

Indicates whether utterances entered by the user should be sent to Amazon Comprehend for sentiment analysis.

createBotVersionResponse_locale :: Lens' CreateBotVersionResponse (Maybe Locale)

Specifies the target locale for the bot.

createBotVersionResponse_createdDate :: Lens' CreateBotVersionResponse (Maybe UTCTime)

The date when the bot version was created.

createBotVersionResponse_name :: Lens' CreateBotVersionResponse (Maybe Text)

The name of the bot.

createBotVersionResponse_version :: Lens' CreateBotVersionResponse (Maybe Text)

The version of the bot.

createBotVersionResponse_idleSessionTTLInSeconds :: Lens' CreateBotVersionResponse (Maybe Natural)

The maximum time in seconds that Amazon Lex retains the data gathered in a conversation. For more information, see PutBot.

createBotVersionResponse_clarificationPrompt :: Lens' CreateBotVersionResponse (Maybe Prompt)

The message that Amazon Lex uses when it doesn't understand the user's request. For more information, see PutBot.

createBotVersionResponse_voiceId :: Lens' CreateBotVersionResponse (Maybe Text)

The Amazon Polly voice ID that Amazon Lex uses for voice interactions with the user.

createBotVersionResponse_lastUpdatedDate :: Lens' CreateBotVersionResponse (Maybe UTCTime)

The date when the $LATEST version of this bot was updated.

createBotVersionResponse_childDirected :: Lens' CreateBotVersionResponse (Maybe Bool)

For each Amazon Lex bot created with the Amazon Lex Model Building Service, you must specify whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to the Children's Online Privacy Protection Act (COPPA) by specifying true or false in the childDirected field. By specifying true in the childDirected field, you confirm that your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. By specifying false in the childDirected field, you confirm that your use of Amazon Lex is not related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. You may not specify a default value for the childDirected field that does not accurately reflect whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA.

If your use of Amazon Lex relates to a website, program, or other application that is directed in whole or in part, to children under age 13, you must obtain any required verifiable parental consent under COPPA. For information regarding the use of Amazon Lex in connection with websites, programs, or other applications that are directed or targeted, in whole or in part, to children under age 13, see the Amazon Lex FAQ.

createBotVersionResponse_description :: Lens' CreateBotVersionResponse (Maybe Text)

A description of the bot.

createBotVersionResponse_httpStatus :: Lens' CreateBotVersionResponse Int

The response's http status code.