Network.AWS.LexModels.CreateIntentVersion

Description

Creates a new version of an intent based on the $LATEST version of the intent. If the $LATEST version of this intent hasn't changed since you last updated it, Amazon Lex doesn't create a new version. It returns the last version you created.

You can update only the $LATEST version of the intent. You can't update the numbered versions that you create with the CreateIntentVersion operation.

When you create a version of an intent, Amazon Lex sets the version to 1. Subsequent versions increment by 1. For more information, see versioning-intro.

This operation requires permissions to perform the lex:CreateIntentVersion action.

Creating a Request

data CreateIntentVersion

See: newCreateIntentVersion smart constructor.

Constructors

CreateIntentVersion'

Fields

checksum :: Maybe Text

Checksum of the $LATEST version of the intent that should be used to create the new version. If you specify a checksum and the $LATEST version of the intent has a different checksum, Amazon Lex returns a PreconditionFailedException exception and doesn't publish a new version. If you don't specify a checksum, Amazon Lex publishes the $LATEST version.

name :: Text

The name of the intent that you want to create a new version of. The name is case sensitive.

newCreateIntentVersion

Arguments

:: Text
$sel:name:CreateIntentVersion'

-> CreateIntentVersion

Create a value of CreateIntentVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checksum:CreateIntentVersion', createIntentVersion_checksum - Checksum of the $LATEST version of the intent that should be used to create the new version. If you specify a checksum and the $LATEST version of the intent has a different checksum, Amazon Lex returns a PreconditionFailedException exception and doesn't publish a new version. If you don't specify a checksum, Amazon Lex publishes the $LATEST version.

$sel:name:CreateIntentVersion', createIntentVersion_name - The name of the intent that you want to create a new version of. The name is case sensitive.

Request Lenses

createIntentVersion_checksum :: Lens' CreateIntentVersion (Maybe Text)

Checksum of the $LATEST version of the intent that should be used to create the new version. If you specify a checksum and the $LATEST version of the intent has a different checksum, Amazon Lex returns a PreconditionFailedException exception and doesn't publish a new version. If you don't specify a checksum, Amazon Lex publishes the $LATEST version.

createIntentVersion_name :: Lens' CreateIntentVersion Text

The name of the intent that you want to create a new version of. The name is case sensitive.

Destructuring the Response

data CreateIntentVersionResponse

See: newCreateIntentVersionResponse smart constructor.

Constructors

CreateIntentVersionResponse'

Fields

fulfillmentActivity :: Maybe FulfillmentActivity

Describes how the intent is fulfilled.

slots :: Maybe [Slot]

An array of slot types that defines the information required to fulfill the intent.

rejectionStatement :: Maybe Statement

If the user answers "no" to the question defined in confirmationPrompt, Amazon Lex responds with this statement to acknowledge that the intent was canceled.

checksum :: Maybe Text

Checksum of the intent version created.

conclusionStatement :: Maybe Statement

After the Lambda function specified in the fulfillmentActivity field fulfills the intent, Amazon Lex conveys this statement to the user.

sampleUtterances :: Maybe [Text]

An array of sample utterances configured for the intent.

parentIntentSignature :: Maybe Text

A unique identifier for a built-in intent.

createdDate :: Maybe POSIX

The date that the intent was created.

kendraConfiguration :: Maybe KendraConfiguration

Configuration information, if any, for connecting an Amazon Kendra index with the AMAZON.KendraSearchIntent intent.

name :: Maybe Text

The name of the intent.

version :: Maybe Text

The version number assigned to the new version of the intent.

inputContexts :: Maybe [InputContext]

An array of InputContext objects that lists the contexts that must be active for Amazon Lex to choose the intent in a conversation with the user.

followUpPrompt :: Maybe FollowUpPrompt

If defined, Amazon Lex uses this prompt to solicit additional user activity after the intent is fulfilled.

lastUpdatedDate :: Maybe POSIX

The date that the intent was updated.

outputContexts :: Maybe [OutputContext]

An array of OutputContext objects that lists the contexts that the intent activates when the intent is fulfilled.

confirmationPrompt :: Maybe Prompt

If defined, the prompt that Amazon Lex uses to confirm the user's intent before fulfilling it.

dialogCodeHook :: Maybe CodeHook

If defined, Amazon Lex invokes this Lambda function for each user input.

description :: Maybe Text

A description of the intent.

httpStatus :: Int

The response's http status code.

newCreateIntentVersionResponse

Arguments

:: Int
$sel:httpStatus:CreateIntentVersionResponse'

-> CreateIntentVersionResponse

Create a value of CreateIntentVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fulfillmentActivity:CreateIntentVersionResponse', createIntentVersionResponse_fulfillmentActivity - Describes how the intent is fulfilled.

$sel:slots:CreateIntentVersionResponse', createIntentVersionResponse_slots - An array of slot types that defines the information required to fulfill the intent.

$sel:rejectionStatement:CreateIntentVersionResponse', createIntentVersionResponse_rejectionStatement - If the user answers "no" to the question defined in confirmationPrompt, Amazon Lex responds with this statement to acknowledge that the intent was canceled.

$sel:checksum:CreateIntentVersion', createIntentVersionResponse_checksum - Checksum of the intent version created.

$sel:conclusionStatement:CreateIntentVersionResponse', createIntentVersionResponse_conclusionStatement - After the Lambda function specified in the fulfillmentActivity field fulfills the intent, Amazon Lex conveys this statement to the user.

$sel:sampleUtterances:CreateIntentVersionResponse', createIntentVersionResponse_sampleUtterances - An array of sample utterances configured for the intent.

$sel:parentIntentSignature:CreateIntentVersionResponse', createIntentVersionResponse_parentIntentSignature - A unique identifier for a built-in intent.

$sel:createdDate:CreateIntentVersionResponse', createIntentVersionResponse_createdDate - The date that the intent was created.

$sel:kendraConfiguration:CreateIntentVersionResponse', createIntentVersionResponse_kendraConfiguration - Configuration information, if any, for connecting an Amazon Kendra index with the AMAZON.KendraSearchIntent intent.

$sel:name:CreateIntentVersion', createIntentVersionResponse_name - The name of the intent.

$sel:version:CreateIntentVersionResponse', createIntentVersionResponse_version - The version number assigned to the new version of the intent.

$sel:inputContexts:CreateIntentVersionResponse', createIntentVersionResponse_inputContexts - An array of InputContext objects that lists the contexts that must be active for Amazon Lex to choose the intent in a conversation with the user.

$sel:followUpPrompt:CreateIntentVersionResponse', createIntentVersionResponse_followUpPrompt - If defined, Amazon Lex uses this prompt to solicit additional user activity after the intent is fulfilled.

$sel:lastUpdatedDate:CreateIntentVersionResponse', createIntentVersionResponse_lastUpdatedDate - The date that the intent was updated.

$sel:outputContexts:CreateIntentVersionResponse', createIntentVersionResponse_outputContexts - An array of OutputContext objects that lists the contexts that the intent activates when the intent is fulfilled.

$sel:confirmationPrompt:CreateIntentVersionResponse', createIntentVersionResponse_confirmationPrompt - If defined, the prompt that Amazon Lex uses to confirm the user's intent before fulfilling it.

$sel:dialogCodeHook:CreateIntentVersionResponse', createIntentVersionResponse_dialogCodeHook - If defined, Amazon Lex invokes this Lambda function for each user input.

$sel:description:CreateIntentVersionResponse', createIntentVersionResponse_description - A description of the intent.

$sel:httpStatus:CreateIntentVersionResponse', createIntentVersionResponse_httpStatus - The response's http status code.

Response Lenses

createIntentVersionResponse_fulfillmentActivity :: Lens' CreateIntentVersionResponse (Maybe FulfillmentActivity)

Describes how the intent is fulfilled.

createIntentVersionResponse_slots :: Lens' CreateIntentVersionResponse (Maybe [Slot])

An array of slot types that defines the information required to fulfill the intent.

createIntentVersionResponse_rejectionStatement :: Lens' CreateIntentVersionResponse (Maybe Statement)

If the user answers "no" to the question defined in confirmationPrompt, Amazon Lex responds with this statement to acknowledge that the intent was canceled.

createIntentVersionResponse_checksum :: Lens' CreateIntentVersionResponse (Maybe Text)

Checksum of the intent version created.

createIntentVersionResponse_conclusionStatement :: Lens' CreateIntentVersionResponse (Maybe Statement)

After the Lambda function specified in the fulfillmentActivity field fulfills the intent, Amazon Lex conveys this statement to the user.

createIntentVersionResponse_sampleUtterances :: Lens' CreateIntentVersionResponse (Maybe [Text])

An array of sample utterances configured for the intent.

createIntentVersionResponse_parentIntentSignature :: Lens' CreateIntentVersionResponse (Maybe Text)

A unique identifier for a built-in intent.

createIntentVersionResponse_createdDate :: Lens' CreateIntentVersionResponse (Maybe UTCTime)

The date that the intent was created.

createIntentVersionResponse_kendraConfiguration :: Lens' CreateIntentVersionResponse (Maybe KendraConfiguration)

Configuration information, if any, for connecting an Amazon Kendra index with the AMAZON.KendraSearchIntent intent.

createIntentVersionResponse_name :: Lens' CreateIntentVersionResponse (Maybe Text)

The name of the intent.

createIntentVersionResponse_version :: Lens' CreateIntentVersionResponse (Maybe Text)

The version number assigned to the new version of the intent.

createIntentVersionResponse_inputContexts :: Lens' CreateIntentVersionResponse (Maybe [InputContext])

An array of InputContext objects that lists the contexts that must be active for Amazon Lex to choose the intent in a conversation with the user.

createIntentVersionResponse_followUpPrompt :: Lens' CreateIntentVersionResponse (Maybe FollowUpPrompt)

If defined, Amazon Lex uses this prompt to solicit additional user activity after the intent is fulfilled.

createIntentVersionResponse_lastUpdatedDate :: Lens' CreateIntentVersionResponse (Maybe UTCTime)

The date that the intent was updated.

createIntentVersionResponse_outputContexts :: Lens' CreateIntentVersionResponse (Maybe [OutputContext])

An array of OutputContext objects that lists the contexts that the intent activates when the intent is fulfilled.

createIntentVersionResponse_confirmationPrompt :: Lens' CreateIntentVersionResponse (Maybe Prompt)

If defined, the prompt that Amazon Lex uses to confirm the user's intent before fulfilling it.

createIntentVersionResponse_dialogCodeHook :: Lens' CreateIntentVersionResponse (Maybe CodeHook)

If defined, Amazon Lex invokes this Lambda function for each user input.

createIntentVersionResponse_description :: Lens' CreateIntentVersionResponse (Maybe Text)

A description of the intent.

createIntentVersionResponse_httpStatus :: Lens' CreateIntentVersionResponse Int

The response's http status code.