Network.AWS.LexModels.CreateSlotTypeVersion

Description

Creates a new version of a slot type based on the $LATEST version of the specified slot type. If the $LATEST version of this resource has not changed since the last version that you created, Amazon Lex doesn't create a new version. It returns the last version that you created.

You can update only the $LATEST version of a slot type. You can't update the numbered versions that you create with the CreateSlotTypeVersion operation.

When you create a version of a slot type, Amazon Lex sets the version to 1. Subsequent versions increment by 1. For more information, see versioning-intro.

This operation requires permissions for the lex:CreateSlotTypeVersion action.

Creating a Request

data CreateSlotTypeVersion

See: newCreateSlotTypeVersion smart constructor.

Constructors

CreateSlotTypeVersion'

Fields

checksum :: Maybe Text

Checksum for the $LATEST version of the slot type that you want to publish. If you specify a checksum and the $LATEST version of the slot type has a different checksum, Amazon Lex returns a PreconditionFailedException exception and doesn't publish the new version. If you don't specify a checksum, Amazon Lex publishes the $LATEST version.

name :: Text

The name of the slot type that you want to create a new version for. The name is case sensitive.

newCreateSlotTypeVersion

Arguments

:: Text
$sel:name:CreateSlotTypeVersion'

-> CreateSlotTypeVersion

Create a value of CreateSlotTypeVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checksum:CreateSlotTypeVersion', createSlotTypeVersion_checksum - Checksum for the $LATEST version of the slot type that you want to publish. If you specify a checksum and the $LATEST version of the slot type has a different checksum, Amazon Lex returns a PreconditionFailedException exception and doesn't publish the new version. If you don't specify a checksum, Amazon Lex publishes the $LATEST version.

$sel:name:CreateSlotTypeVersion', createSlotTypeVersion_name - The name of the slot type that you want to create a new version for. The name is case sensitive.

Request Lenses

createSlotTypeVersion_checksum :: Lens' CreateSlotTypeVersion (Maybe Text)

Checksum for the $LATEST version of the slot type that you want to publish. If you specify a checksum and the $LATEST version of the slot type has a different checksum, Amazon Lex returns a PreconditionFailedException exception and doesn't publish the new version. If you don't specify a checksum, Amazon Lex publishes the $LATEST version.

createSlotTypeVersion_name :: Lens' CreateSlotTypeVersion Text

The name of the slot type that you want to create a new version for. The name is case sensitive.

Destructuring the Response

data CreateSlotTypeVersionResponse

See: newCreateSlotTypeVersionResponse smart constructor.

Constructors

CreateSlotTypeVersionResponse'

Fields

parentSlotTypeSignature :: Maybe Text

The built-in slot type used a the parent of the slot type.

slotTypeConfigurations :: Maybe [SlotTypeConfiguration]

Configuration information that extends the parent built-in slot type.

checksum :: Maybe Text

Checksum of the $LATEST version of the slot type.

valueSelectionStrategy :: Maybe SlotValueSelectionStrategy

The strategy that Amazon Lex uses to determine the value of the slot. For more information, see PutSlotType.

createdDate :: Maybe POSIX

The date that the slot type was created.

name :: Maybe Text

The name of the slot type.

version :: Maybe Text

The version assigned to the new slot type version.

lastUpdatedDate :: Maybe POSIX

The date that the slot type was updated. When you create a resource, the creation date and last update date are the same.

description :: Maybe Text

A description of the slot type.

enumerationValues :: Maybe [EnumerationValue]

A list of EnumerationValue objects that defines the values that the slot type can take.

httpStatus :: Int

The response's http status code.

newCreateSlotTypeVersionResponse

Arguments

:: Int
$sel:httpStatus:CreateSlotTypeVersionResponse'

-> CreateSlotTypeVersionResponse

Create a value of CreateSlotTypeVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parentSlotTypeSignature:CreateSlotTypeVersionResponse', createSlotTypeVersionResponse_parentSlotTypeSignature - The built-in slot type used a the parent of the slot type.

$sel:slotTypeConfigurations:CreateSlotTypeVersionResponse', createSlotTypeVersionResponse_slotTypeConfigurations - Configuration information that extends the parent built-in slot type.

$sel:checksum:CreateSlotTypeVersion', createSlotTypeVersionResponse_checksum - Checksum of the $LATEST version of the slot type.

$sel:valueSelectionStrategy:CreateSlotTypeVersionResponse', createSlotTypeVersionResponse_valueSelectionStrategy - The strategy that Amazon Lex uses to determine the value of the slot. For more information, see PutSlotType.

$sel:createdDate:CreateSlotTypeVersionResponse', createSlotTypeVersionResponse_createdDate - The date that the slot type was created.

$sel:name:CreateSlotTypeVersion', createSlotTypeVersionResponse_name - The name of the slot type.

$sel:version:CreateSlotTypeVersionResponse', createSlotTypeVersionResponse_version - The version assigned to the new slot type version.

$sel:lastUpdatedDate:CreateSlotTypeVersionResponse', createSlotTypeVersionResponse_lastUpdatedDate - The date that the slot type was updated. When you create a resource, the creation date and last update date are the same.

$sel:description:CreateSlotTypeVersionResponse', createSlotTypeVersionResponse_description - A description of the slot type.

$sel:enumerationValues:CreateSlotTypeVersionResponse', createSlotTypeVersionResponse_enumerationValues - A list of EnumerationValue objects that defines the values that the slot type can take.

$sel:httpStatus:CreateSlotTypeVersionResponse', createSlotTypeVersionResponse_httpStatus - The response's http status code.

Response Lenses

createSlotTypeVersionResponse_parentSlotTypeSignature :: Lens' CreateSlotTypeVersionResponse (Maybe Text)

The built-in slot type used a the parent of the slot type.

createSlotTypeVersionResponse_slotTypeConfigurations :: Lens' CreateSlotTypeVersionResponse (Maybe [SlotTypeConfiguration])

Configuration information that extends the parent built-in slot type.

createSlotTypeVersionResponse_checksum :: Lens' CreateSlotTypeVersionResponse (Maybe Text)

Checksum of the $LATEST version of the slot type.

createSlotTypeVersionResponse_valueSelectionStrategy :: Lens' CreateSlotTypeVersionResponse (Maybe SlotValueSelectionStrategy)

The strategy that Amazon Lex uses to determine the value of the slot. For more information, see PutSlotType.

createSlotTypeVersionResponse_createdDate :: Lens' CreateSlotTypeVersionResponse (Maybe UTCTime)

The date that the slot type was created.

createSlotTypeVersionResponse_name :: Lens' CreateSlotTypeVersionResponse (Maybe Text)

The name of the slot type.

createSlotTypeVersionResponse_version :: Lens' CreateSlotTypeVersionResponse (Maybe Text)

The version assigned to the new slot type version.

createSlotTypeVersionResponse_lastUpdatedDate :: Lens' CreateSlotTypeVersionResponse (Maybe UTCTime)

The date that the slot type was updated. When you create a resource, the creation date and last update date are the same.

createSlotTypeVersionResponse_description :: Lens' CreateSlotTypeVersionResponse (Maybe Text)

A description of the slot type.

createSlotTypeVersionResponse_enumerationValues :: Lens' CreateSlotTypeVersionResponse (Maybe [EnumerationValue])

A list of EnumerationValue objects that defines the values that the slot type can take.

createSlotTypeVersionResponse_httpStatus :: Lens' CreateSlotTypeVersionResponse Int

The response's http status code.