Network.AWS.LexModels.DeleteBot

Description

Deletes all versions of the bot, including the $LATEST version. To delete a specific version of the bot, use the DeleteBotVersion operation. The DeleteBot operation doesn't immediately remove the bot schema. Instead, it is marked for deletion and removed later.

Amazon Lex stores utterances indefinitely for improving the ability of your bot to respond to user inputs. These utterances are not removed when the bot is deleted. To remove the utterances, use the DeleteUtterances operation.

If a bot has an alias, you can't delete it. Instead, the DeleteBot operation returns a ResourceInUseException exception that includes a reference to the alias that refers to the bot. To remove the reference to the bot, delete the alias. If you get the same exception again, delete the referring alias until the DeleteBot operation is successful.

This operation requires permissions for the lex:DeleteBot action.

Creating a Request

data DeleteBot

See: newDeleteBot smart constructor.

Constructors

DeleteBot'

Fields

name :: Text

The name of the bot. The name is case sensitive.

newDeleteBot

Arguments

:: Text
$sel:name:DeleteBot'

-> DeleteBot

Create a value of DeleteBot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DeleteBot', deleteBot_name - The name of the bot. The name is case sensitive.

Request Lenses

deleteBot_name :: Lens' DeleteBot Text

The name of the bot. The name is case sensitive.

Destructuring the Response

data DeleteBotResponse

See: newDeleteBotResponse smart constructor.

Constructors

DeleteBotResponse'

newDeleteBotResponse :: DeleteBotResponse

Create a value of DeleteBotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.