Network.AWS.LexModels.DeleteIntent

Description

Deletes all versions of the intent, including the $LATEST version. To delete a specific version of the intent, use the DeleteIntentVersion operation.

You can delete a version of an intent only if it is not referenced. To delete an intent that is referred to in one or more bots (see how-it-works), you must remove those references first.

If you get the ResourceInUseException exception, it provides an example reference that shows where the intent is referenced. To remove the reference to the intent, either update the bot or delete it. If you get the same exception when you attempt to delete the intent again, repeat until the intent has no references and the call to DeleteIntent is successful.

This operation requires permission for the lex:DeleteIntent action.

Creating a Request

data DeleteIntent

See: newDeleteIntent smart constructor.

Constructors

DeleteIntent'

Fields

name :: Text

The name of the intent. The name is case sensitive.

newDeleteIntent

Arguments

:: Text
$sel:name:DeleteIntent'

-> DeleteIntent

Create a value of DeleteIntent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DeleteIntent', deleteIntent_name - The name of the intent. The name is case sensitive.

Request Lenses

deleteIntent_name :: Lens' DeleteIntent Text

The name of the intent. The name is case sensitive.

Destructuring the Response

data DeleteIntentResponse

See: newDeleteIntentResponse smart constructor.

Constructors

DeleteIntentResponse'

newDeleteIntentResponse :: DeleteIntentResponse

Create a value of DeleteIntentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.