Network.AWS.LexModels.DeleteIntentVersion

Description

Deletes a specific version of an intent. To delete all versions of a intent, use the DeleteIntent operation.

This operation requires permissions for the lex:DeleteIntentVersion action.

Creating a Request

data DeleteIntentVersion

See: newDeleteIntentVersion smart constructor.

Constructors

DeleteIntentVersion'

Fields

name :: Text

The name of the intent.

version :: Text

The version of the intent to delete. You cannot delete the $LATEST version of the intent. To delete the $LATEST version, use the DeleteIntent operation.

newDeleteIntentVersion

Arguments

:: Text
$sel:name:DeleteIntentVersion'

-> Text
$sel:version:DeleteIntentVersion'

-> DeleteIntentVersion

Create a value of DeleteIntentVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DeleteIntentVersion', deleteIntentVersion_name - The name of the intent.

$sel:version:DeleteIntentVersion', deleteIntentVersion_version - The version of the intent to delete. You cannot delete the $LATEST version of the intent. To delete the $LATEST version, use the DeleteIntent operation.

Request Lenses

deleteIntentVersion_name :: Lens' DeleteIntentVersion Text

The name of the intent.

deleteIntentVersion_version :: Lens' DeleteIntentVersion Text

The version of the intent to delete. You cannot delete the $LATEST version of the intent. To delete the $LATEST version, use the DeleteIntent operation.

Destructuring the Response

data DeleteIntentVersionResponse

See: newDeleteIntentVersionResponse smart constructor.

Constructors

DeleteIntentVersionResponse'

newDeleteIntentVersionResponse :: DeleteIntentVersionResponse

Create a value of DeleteIntentVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.