Network.AWS.LexModels.DeleteSlotTypeVersion

Description

Deletes a specific version of a slot type. To delete all versions of a slot type, use the DeleteSlotType operation.

This operation requires permissions for the lex:DeleteSlotTypeVersion action.

Creating a Request

data DeleteSlotTypeVersion

See: newDeleteSlotTypeVersion smart constructor.

Constructors

DeleteSlotTypeVersion'

Fields

name :: Text

The name of the slot type.

version :: Text

The version of the slot type to delete. You cannot delete the $LATEST version of the slot type. To delete the $LATEST version, use the DeleteSlotType operation.

newDeleteSlotTypeVersion

Arguments

:: Text
$sel:name:DeleteSlotTypeVersion'

-> Text
$sel:version:DeleteSlotTypeVersion'

-> DeleteSlotTypeVersion

Create a value of DeleteSlotTypeVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DeleteSlotTypeVersion', deleteSlotTypeVersion_name - The name of the slot type.

$sel:version:DeleteSlotTypeVersion', deleteSlotTypeVersion_version - The version of the slot type to delete. You cannot delete the $LATEST version of the slot type. To delete the $LATEST version, use the DeleteSlotType operation.

Request Lenses

deleteSlotTypeVersion_name :: Lens' DeleteSlotTypeVersion Text

The name of the slot type.

deleteSlotTypeVersion_version :: Lens' DeleteSlotTypeVersion Text

The version of the slot type to delete. You cannot delete the $LATEST version of the slot type. To delete the $LATEST version, use the DeleteSlotType operation.

Destructuring the Response

data DeleteSlotTypeVersionResponse

See: newDeleteSlotTypeVersionResponse smart constructor.

Constructors

DeleteSlotTypeVersionResponse'

newDeleteSlotTypeVersionResponse :: DeleteSlotTypeVersionResponse

Create a value of DeleteSlotTypeVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.