Network.AWS.LexModels.GetBotAlias

Description

Returns information about an Amazon Lex bot alias. For more information about aliases, see versioning-aliases.

This operation requires permissions for the lex:GetBotAlias action.

Creating a Request

data GetBotAlias

See: newGetBotAlias smart constructor.

Constructors

GetBotAlias'

Fields

name :: Text

The name of the bot alias. The name is case sensitive.

botName :: Text

The name of the bot.

newGetBotAlias

Arguments

:: Text
$sel:name:GetBotAlias'

-> Text
$sel:botName:GetBotAlias'

-> GetBotAlias

Create a value of GetBotAlias with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetBotAlias', getBotAlias_name - The name of the bot alias. The name is case sensitive.

$sel:botName:GetBotAlias', getBotAlias_botName - The name of the bot.

Request Lenses

getBotAlias_name :: Lens' GetBotAlias Text

The name of the bot alias. The name is case sensitive.

getBotAlias_botName :: Lens' GetBotAlias Text

The name of the bot.

Destructuring the Response

data GetBotAliasResponse

See: newGetBotAliasResponse smart constructor.

Constructors

GetBotAliasResponse'

Fields

checksum :: Maybe Text

Checksum of the bot alias.

botVersion :: Maybe Text

The version of the bot that the alias points to.

botName :: Maybe Text

The name of the bot that the alias points to.

createdDate :: Maybe POSIX

The date that the bot alias was created.

name :: Maybe Text

The name of the bot alias.

conversationLogs :: Maybe ConversationLogsResponse

The settings that determine how Amazon Lex uses conversation logs for the alias.

lastUpdatedDate :: Maybe POSIX

The date that the bot alias was updated. When you create a resource, the creation date and the last updated date are the same.

description :: Maybe Text

A description of the bot alias.

httpStatus :: Int

The response's http status code.

newGetBotAliasResponse

Arguments

:: Int
$sel:httpStatus:GetBotAliasResponse'

-> GetBotAliasResponse

Create a value of GetBotAliasResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checksum:GetBotAliasResponse', getBotAliasResponse_checksum - Checksum of the bot alias.

$sel:botVersion:GetBotAliasResponse', getBotAliasResponse_botVersion - The version of the bot that the alias points to.

$sel:botName:GetBotAlias', getBotAliasResponse_botName - The name of the bot that the alias points to.

$sel:createdDate:GetBotAliasResponse', getBotAliasResponse_createdDate - The date that the bot alias was created.

$sel:name:GetBotAlias', getBotAliasResponse_name - The name of the bot alias.

$sel:conversationLogs:GetBotAliasResponse', getBotAliasResponse_conversationLogs - The settings that determine how Amazon Lex uses conversation logs for the alias.

$sel:lastUpdatedDate:GetBotAliasResponse', getBotAliasResponse_lastUpdatedDate - The date that the bot alias was updated. When you create a resource, the creation date and the last updated date are the same.

$sel:description:GetBotAliasResponse', getBotAliasResponse_description - A description of the bot alias.

$sel:httpStatus:GetBotAliasResponse', getBotAliasResponse_httpStatus - The response's http status code.

Response Lenses

getBotAliasResponse_checksum :: Lens' GetBotAliasResponse (Maybe Text)

Checksum of the bot alias.

getBotAliasResponse_botVersion :: Lens' GetBotAliasResponse (Maybe Text)

The version of the bot that the alias points to.

getBotAliasResponse_botName :: Lens' GetBotAliasResponse (Maybe Text)

The name of the bot that the alias points to.

getBotAliasResponse_createdDate :: Lens' GetBotAliasResponse (Maybe UTCTime)

The date that the bot alias was created.

getBotAliasResponse_name :: Lens' GetBotAliasResponse (Maybe Text)

The name of the bot alias.

getBotAliasResponse_conversationLogs :: Lens' GetBotAliasResponse (Maybe ConversationLogsResponse)

The settings that determine how Amazon Lex uses conversation logs for the alias.

getBotAliasResponse_lastUpdatedDate :: Lens' GetBotAliasResponse (Maybe UTCTime)

The date that the bot alias was updated. When you create a resource, the creation date and the last updated date are the same.

getBotAliasResponse_description :: Lens' GetBotAliasResponse (Maybe Text)

A description of the bot alias.

getBotAliasResponse_httpStatus :: Lens' GetBotAliasResponse Int

The response's http status code.