Network.AWS.LexModels.GetBotAliases

Description

Returns a list of aliases for a specified Amazon Lex bot.

This operation requires permissions for the lex:GetBotAliases action.

This operation returns paginated results.

Creating a Request

data GetBotAliases

See: newGetBotAliases smart constructor.

Constructors

GetBotAliases'

Fields

nameContains :: Maybe Text

Substring to match in bot alias names. An alias will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

nextToken :: Maybe Text

A pagination token for fetching the next page of aliases. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of aliases, specify the pagination token in the next request.

maxResults :: Maybe Natural

The maximum number of aliases to return in the response. The default is 50. .

botName :: Text

The name of the bot.

newGetBotAliases

Arguments

:: Text
$sel:botName:GetBotAliases'

-> GetBotAliases

Create a value of GetBotAliases with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nameContains:GetBotAliases', getBotAliases_nameContains - Substring to match in bot alias names. An alias will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

$sel:nextToken:GetBotAliases', getBotAliases_nextToken - A pagination token for fetching the next page of aliases. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of aliases, specify the pagination token in the next request.

$sel:maxResults:GetBotAliases', getBotAliases_maxResults - The maximum number of aliases to return in the response. The default is 50. .

$sel:botName:GetBotAliases', getBotAliases_botName - The name of the bot.

Request Lenses

getBotAliases_nameContains :: Lens' GetBotAliases (Maybe Text)

Substring to match in bot alias names. An alias will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

getBotAliases_nextToken :: Lens' GetBotAliases (Maybe Text)

A pagination token for fetching the next page of aliases. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of aliases, specify the pagination token in the next request.

getBotAliases_maxResults :: Lens' GetBotAliases (Maybe Natural)

The maximum number of aliases to return in the response. The default is 50. .

getBotAliases_botName :: Lens' GetBotAliases Text

The name of the bot.

Destructuring the Response

data GetBotAliasesResponse

See: newGetBotAliasesResponse smart constructor.

Constructors

GetBotAliasesResponse'

Fields

nextToken :: Maybe Text

A pagination token for fetching next page of aliases. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of aliases, specify the pagination token in the next request.

botAliases :: Maybe [BotAliasMetadata]

An array of BotAliasMetadata objects, each describing a bot alias.

httpStatus :: Int

The response's http status code.

newGetBotAliasesResponse

Arguments

:: Int
$sel:httpStatus:GetBotAliasesResponse'

-> GetBotAliasesResponse

Create a value of GetBotAliasesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetBotAliases', getBotAliasesResponse_nextToken - A pagination token for fetching next page of aliases. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of aliases, specify the pagination token in the next request.

$sel:botAliases:GetBotAliasesResponse', getBotAliasesResponse_botAliases - An array of BotAliasMetadata objects, each describing a bot alias.

$sel:httpStatus:GetBotAliasesResponse', getBotAliasesResponse_httpStatus - The response's http status code.

Response Lenses

getBotAliasesResponse_nextToken :: Lens' GetBotAliasesResponse (Maybe Text)

A pagination token for fetching next page of aliases. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of aliases, specify the pagination token in the next request.

getBotAliasesResponse_botAliases :: Lens' GetBotAliasesResponse (Maybe [BotAliasMetadata])

An array of BotAliasMetadata objects, each describing a bot alias.

getBotAliasesResponse_httpStatus :: Lens' GetBotAliasesResponse Int

The response's http status code.