Network.AWS.LexModels.GetBotChannelAssociation

Description

Returns information about the association between an Amazon Lex bot and a messaging platform.

This operation requires permissions for the lex:GetBotChannelAssociation action.

Creating a Request

data GetBotChannelAssociation

See: newGetBotChannelAssociation smart constructor.

Constructors

GetBotChannelAssociation'

Fields

name :: Text

The name of the association between the bot and the channel. The name is case sensitive.

botName :: Text

The name of the Amazon Lex bot.

botAlias :: Text

An alias pointing to the specific version of the Amazon Lex bot to which this association is being made.

newGetBotChannelAssociation

Arguments

:: Text
$sel:name:GetBotChannelAssociation'

-> Text
$sel:botName:GetBotChannelAssociation'

-> Text
$sel:botAlias:GetBotChannelAssociation'

-> GetBotChannelAssociation

Create a value of GetBotChannelAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetBotChannelAssociation', getBotChannelAssociation_name - The name of the association between the bot and the channel. The name is case sensitive.

$sel:botName:GetBotChannelAssociation', getBotChannelAssociation_botName - The name of the Amazon Lex bot.

$sel:botAlias:GetBotChannelAssociation', getBotChannelAssociation_botAlias - An alias pointing to the specific version of the Amazon Lex bot to which this association is being made.

Request Lenses

getBotChannelAssociation_name :: Lens' GetBotChannelAssociation Text

The name of the association between the bot and the channel. The name is case sensitive.

getBotChannelAssociation_botName :: Lens' GetBotChannelAssociation Text

The name of the Amazon Lex bot.

getBotChannelAssociation_botAlias :: Lens' GetBotChannelAssociation Text

An alias pointing to the specific version of the Amazon Lex bot to which this association is being made.

Destructuring the Response

data GetBotChannelAssociationResponse

See: newGetBotChannelAssociationResponse smart constructor.

Constructors

GetBotChannelAssociationResponse'

Fields

botAlias :: Maybe Text

An alias pointing to the specific version of the Amazon Lex bot to which this association is being made.

createdDate :: Maybe POSIX

The date that the association between the bot and the channel was created.

status :: Maybe ChannelStatus

The status of the bot channel.

CREATED - The channel has been created and is ready for use.

IN_PROGRESS - Channel creation is in progress.

FAILED - There was an error creating the channel. For information about the reason for the failure, see the failureReason field.

botConfiguration :: Maybe (Sensitive (HashMap Text Text))

Provides information that the messaging platform needs to communicate with the Amazon Lex bot.

botName :: Maybe Text

The name of the Amazon Lex bot.

name :: Maybe Text

The name of the association between the bot and the channel.

failureReason :: Maybe Text

If status is FAILED, Amazon Lex provides the reason that it failed to create the association.

description :: Maybe Text

A description of the association between the bot and the channel.

type' :: Maybe ChannelType

The type of the messaging platform.

httpStatus :: Int

The response's http status code.

newGetBotChannelAssociationResponse

Arguments

:: Int
$sel:httpStatus:GetBotChannelAssociationResponse'

-> GetBotChannelAssociationResponse

Create a value of GetBotChannelAssociationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:botAlias:GetBotChannelAssociation', getBotChannelAssociationResponse_botAlias - An alias pointing to the specific version of the Amazon Lex bot to which this association is being made.

$sel:createdDate:GetBotChannelAssociationResponse', getBotChannelAssociationResponse_createdDate - The date that the association between the bot and the channel was created.

$sel:status:GetBotChannelAssociationResponse', getBotChannelAssociationResponse_status - The status of the bot channel.

CREATED - The channel has been created and is ready for use.

IN_PROGRESS - Channel creation is in progress.

FAILED - There was an error creating the channel. For information about the reason for the failure, see the failureReason field.

$sel:botConfiguration:GetBotChannelAssociationResponse', getBotChannelAssociationResponse_botConfiguration - Provides information that the messaging platform needs to communicate with the Amazon Lex bot.

$sel:botName:GetBotChannelAssociation', getBotChannelAssociationResponse_botName - The name of the Amazon Lex bot.

$sel:name:GetBotChannelAssociation', getBotChannelAssociationResponse_name - The name of the association between the bot and the channel.

$sel:failureReason:GetBotChannelAssociationResponse', getBotChannelAssociationResponse_failureReason - If status is FAILED, Amazon Lex provides the reason that it failed to create the association.

$sel:description:GetBotChannelAssociationResponse', getBotChannelAssociationResponse_description - A description of the association between the bot and the channel.

$sel:type':GetBotChannelAssociationResponse', getBotChannelAssociationResponse_type - The type of the messaging platform.

$sel:httpStatus:GetBotChannelAssociationResponse', getBotChannelAssociationResponse_httpStatus - The response's http status code.

Response Lenses

getBotChannelAssociationResponse_botAlias :: Lens' GetBotChannelAssociationResponse (Maybe Text)

An alias pointing to the specific version of the Amazon Lex bot to which this association is being made.

getBotChannelAssociationResponse_createdDate :: Lens' GetBotChannelAssociationResponse (Maybe UTCTime)

The date that the association between the bot and the channel was created.

getBotChannelAssociationResponse_status :: Lens' GetBotChannelAssociationResponse (Maybe ChannelStatus)

The status of the bot channel.

CREATED - The channel has been created and is ready for use.

IN_PROGRESS - Channel creation is in progress.

FAILED - There was an error creating the channel. For information about the reason for the failure, see the failureReason field.

getBotChannelAssociationResponse_botConfiguration :: Lens' GetBotChannelAssociationResponse (Maybe (HashMap Text Text))

Provides information that the messaging platform needs to communicate with the Amazon Lex bot.

getBotChannelAssociationResponse_botName :: Lens' GetBotChannelAssociationResponse (Maybe Text)

The name of the Amazon Lex bot.

getBotChannelAssociationResponse_name :: Lens' GetBotChannelAssociationResponse (Maybe Text)

The name of the association between the bot and the channel.

getBotChannelAssociationResponse_failureReason :: Lens' GetBotChannelAssociationResponse (Maybe Text)

If status is FAILED, Amazon Lex provides the reason that it failed to create the association.

getBotChannelAssociationResponse_description :: Lens' GetBotChannelAssociationResponse (Maybe Text)

A description of the association between the bot and the channel.

getBotChannelAssociationResponse_type :: Lens' GetBotChannelAssociationResponse (Maybe ChannelType)

The type of the messaging platform.

getBotChannelAssociationResponse_httpStatus :: Lens' GetBotChannelAssociationResponse Int

The response's http status code.