Network.AWS.LexModels.GetBotVersions

Description

Gets information about all of the versions of a bot.

The GetBotVersions operation returns a BotMetadata object for each version of a bot. For example, if a bot has three numbered versions, the GetBotVersions operation returns four BotMetadata objects in the response, one for each numbered version and one for the $LATEST version.

The GetBotVersions operation always returns at least one version, the $LATEST version.

This operation requires permissions for the lex:GetBotVersions action.

This operation returns paginated results.

Creating a Request

data GetBotVersions

See: newGetBotVersions smart constructor.

Constructors

GetBotVersions'

Fields

nextToken :: Maybe Text

A pagination token for fetching the next page of bot versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

maxResults :: Maybe Natural

The maximum number of bot versions to return in the response. The default is 10.

name :: Text

The name of the bot for which versions should be returned.

newGetBotVersions

Arguments

:: Text
$sel:name:GetBotVersions'

-> GetBotVersions

Create a value of GetBotVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetBotVersions', getBotVersions_nextToken - A pagination token for fetching the next page of bot versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

$sel:maxResults:GetBotVersions', getBotVersions_maxResults - The maximum number of bot versions to return in the response. The default is 10.

$sel:name:GetBotVersions', getBotVersions_name - The name of the bot for which versions should be returned.

Request Lenses

getBotVersions_nextToken :: Lens' GetBotVersions (Maybe Text)

A pagination token for fetching the next page of bot versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

getBotVersions_maxResults :: Lens' GetBotVersions (Maybe Natural)

The maximum number of bot versions to return in the response. The default is 10.

getBotVersions_name :: Lens' GetBotVersions Text

The name of the bot for which versions should be returned.

Destructuring the Response

data GetBotVersionsResponse

See: newGetBotVersionsResponse smart constructor.

Constructors

GetBotVersionsResponse'

Fields

bots :: Maybe [BotMetadata]

An array of BotMetadata objects, one for each numbered version of the bot plus one for the $LATEST version.

nextToken :: Maybe Text

A pagination token for fetching the next page of bot versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

httpStatus :: Int

The response's http status code.

newGetBotVersionsResponse

Arguments

:: Int
$sel:httpStatus:GetBotVersionsResponse'

-> GetBotVersionsResponse

Create a value of GetBotVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bots:GetBotVersionsResponse', getBotVersionsResponse_bots - An array of BotMetadata objects, one for each numbered version of the bot plus one for the $LATEST version.

$sel:nextToken:GetBotVersions', getBotVersionsResponse_nextToken - A pagination token for fetching the next page of bot versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

$sel:httpStatus:GetBotVersionsResponse', getBotVersionsResponse_httpStatus - The response's http status code.

Response Lenses

getBotVersionsResponse_bots :: Lens' GetBotVersionsResponse (Maybe [BotMetadata])

An array of BotMetadata objects, one for each numbered version of the bot plus one for the $LATEST version.

getBotVersionsResponse_nextToken :: Lens' GetBotVersionsResponse (Maybe Text)

A pagination token for fetching the next page of bot versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

getBotVersionsResponse_httpStatus :: Lens' GetBotVersionsResponse Int

The response's http status code.