Network.AWS.LexModels.GetBots

Description

Returns bot information as follows:

If you provide the nameContains field, the response includes information for the $LATEST version of all bots whose name contains the specified string.

If you don't specify the nameContains field, the operation returns information about the $LATEST version of all of your bots.

This operation requires permission for the lex:GetBots action.

This operation returns paginated results.

Creating a Request

data GetBots

See: newGetBots smart constructor.

Constructors

GetBots'

Fields

nameContains :: Maybe Text

Substring to match in bot names. A bot will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

nextToken :: Maybe Text

A pagination token that fetches the next page of bots. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of bots, specify the pagination token in the next request.

maxResults :: Maybe Natural

The maximum number of bots to return in the response that the request will return. The default is 10.

newGetBots :: GetBots

Create a value of GetBots with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nameContains:GetBots', getBots_nameContains - Substring to match in bot names. A bot will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

$sel:nextToken:GetBots', getBots_nextToken - A pagination token that fetches the next page of bots. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of bots, specify the pagination token in the next request.

$sel:maxResults:GetBots', getBots_maxResults - The maximum number of bots to return in the response that the request will return. The default is 10.

Request Lenses

getBots_nameContains :: Lens' GetBots (Maybe Text)

Substring to match in bot names. A bot will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

getBots_nextToken :: Lens' GetBots (Maybe Text)

A pagination token that fetches the next page of bots. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of bots, specify the pagination token in the next request.

getBots_maxResults :: Lens' GetBots (Maybe Natural)

The maximum number of bots to return in the response that the request will return. The default is 10.

Destructuring the Response

data GetBotsResponse

See: newGetBotsResponse smart constructor.

Constructors

GetBotsResponse'

Fields

nextToken :: Maybe Text

If the response is truncated, it includes a pagination token that you can specify in your next request to fetch the next page of bots.

bots :: Maybe [BotMetadata]

An array of botMetadata objects, with one entry for each bot.

httpStatus :: Int

The response's http status code.

newGetBotsResponse

Arguments

:: Int
$sel:httpStatus:GetBotsResponse'

-> GetBotsResponse

Create a value of GetBotsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetBots', getBotsResponse_nextToken - If the response is truncated, it includes a pagination token that you can specify in your next request to fetch the next page of bots.

$sel:bots:GetBotsResponse', getBotsResponse_bots - An array of botMetadata objects, with one entry for each bot.

$sel:httpStatus:GetBotsResponse', getBotsResponse_httpStatus - The response's http status code.

Response Lenses

getBotsResponse_nextToken :: Lens' GetBotsResponse (Maybe Text)

If the response is truncated, it includes a pagination token that you can specify in your next request to fetch the next page of bots.

getBotsResponse_bots :: Lens' GetBotsResponse (Maybe [BotMetadata])

An array of botMetadata objects, with one entry for each bot.

getBotsResponse_httpStatus :: Lens' GetBotsResponse Int

The response's http status code.