Network.AWS.LexModels.GetBuiltinSlotTypes

Description

Gets a list of built-in slot types that meet the specified criteria.

For a list of built-in slot types, see Slot Type Reference in the Alexa Skills Kit.

This operation requires permission for the lex:GetBuiltInSlotTypes action.

This operation returns paginated results.

Creating a Request

data GetBuiltinSlotTypes

See: newGetBuiltinSlotTypes smart constructor.

Constructors

GetBuiltinSlotTypes'

Fields

signatureContains :: Maybe Text

Substring to match in built-in slot type signatures. A slot type will be returned if any part of its signature matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

nextToken :: Maybe Text

A pagination token that fetches the next page of slot types. If the response to this API call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of slot types, specify the pagination token in the next request.

maxResults :: Maybe Natural

The maximum number of slot types to return in the response. The default is 10.

locale :: Maybe Locale

A list of locales that the slot type supports.

newGetBuiltinSlotTypes :: GetBuiltinSlotTypes

Create a value of GetBuiltinSlotTypes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:signatureContains:GetBuiltinSlotTypes', getBuiltinSlotTypes_signatureContains - Substring to match in built-in slot type signatures. A slot type will be returned if any part of its signature matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

$sel:nextToken:GetBuiltinSlotTypes', getBuiltinSlotTypes_nextToken - A pagination token that fetches the next page of slot types. If the response to this API call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of slot types, specify the pagination token in the next request.

$sel:maxResults:GetBuiltinSlotTypes', getBuiltinSlotTypes_maxResults - The maximum number of slot types to return in the response. The default is 10.

$sel:locale:GetBuiltinSlotTypes', getBuiltinSlotTypes_locale - A list of locales that the slot type supports.

Request Lenses

getBuiltinSlotTypes_signatureContains :: Lens' GetBuiltinSlotTypes (Maybe Text)

Substring to match in built-in slot type signatures. A slot type will be returned if any part of its signature matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

getBuiltinSlotTypes_nextToken :: Lens' GetBuiltinSlotTypes (Maybe Text)

A pagination token that fetches the next page of slot types. If the response to this API call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of slot types, specify the pagination token in the next request.

getBuiltinSlotTypes_maxResults :: Lens' GetBuiltinSlotTypes (Maybe Natural)

The maximum number of slot types to return in the response. The default is 10.

getBuiltinSlotTypes_locale :: Lens' GetBuiltinSlotTypes (Maybe Locale)

A list of locales that the slot type supports.

Destructuring the Response

data GetBuiltinSlotTypesResponse

See: newGetBuiltinSlotTypesResponse smart constructor.

Constructors

GetBuiltinSlotTypesResponse'

Fields

slotTypes :: Maybe [BuiltinSlotTypeMetadata]

An array of BuiltInSlotTypeMetadata objects, one entry for each slot type returned.

nextToken :: Maybe Text

If the response is truncated, the response includes a pagination token that you can use in your next request to fetch the next page of slot types.

httpStatus :: Int

The response's http status code.

newGetBuiltinSlotTypesResponse

Arguments

:: Int
$sel:httpStatus:GetBuiltinSlotTypesResponse'

-> GetBuiltinSlotTypesResponse

Create a value of GetBuiltinSlotTypesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:slotTypes:GetBuiltinSlotTypesResponse', getBuiltinSlotTypesResponse_slotTypes - An array of BuiltInSlotTypeMetadata objects, one entry for each slot type returned.

$sel:nextToken:GetBuiltinSlotTypes', getBuiltinSlotTypesResponse_nextToken - If the response is truncated, the response includes a pagination token that you can use in your next request to fetch the next page of slot types.

$sel:httpStatus:GetBuiltinSlotTypesResponse', getBuiltinSlotTypesResponse_httpStatus - The response's http status code.

Response Lenses

getBuiltinSlotTypesResponse_slotTypes :: Lens' GetBuiltinSlotTypesResponse (Maybe [BuiltinSlotTypeMetadata])

An array of BuiltInSlotTypeMetadata objects, one entry for each slot type returned.

getBuiltinSlotTypesResponse_nextToken :: Lens' GetBuiltinSlotTypesResponse (Maybe Text)

If the response is truncated, the response includes a pagination token that you can use in your next request to fetch the next page of slot types.

getBuiltinSlotTypesResponse_httpStatus :: Lens' GetBuiltinSlotTypesResponse Int

The response's http status code.