Network.AWS.LexModels.GetExport

Description

Exports the contents of a Amazon Lex resource in a specified format.

Creating a Request

data GetExport

See: newGetExport smart constructor.

Constructors

GetExport'

Fields

name :: Text

The name of the bot to export.

version :: Text

The version of the bot to export.

resourceType :: ResourceType

The type of resource to export.

exportType :: ExportType

The format of the exported data.

newGetExport

Arguments

:: Text
$sel:name:GetExport'

-> Text
$sel:version:GetExport'

-> ResourceType
$sel:resourceType:GetExport'

-> ExportType
$sel:exportType:GetExport'

-> GetExport

Create a value of GetExport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetExport', getExport_name - The name of the bot to export.

$sel:version:GetExport', getExport_version - The version of the bot to export.

$sel:resourceType:GetExport', getExport_resourceType - The type of resource to export.

$sel:exportType:GetExport', getExport_exportType - The format of the exported data.

Request Lenses

getExport_name :: Lens' GetExport Text

The name of the bot to export.

getExport_version :: Lens' GetExport Text

The version of the bot to export.

getExport_resourceType :: Lens' GetExport ResourceType

The type of resource to export.

getExport_exportType :: Lens' GetExport ExportType

The format of the exported data.

Destructuring the Response

data GetExportResponse

See: newGetExportResponse smart constructor.

Constructors

GetExportResponse'

Fields

failureReason :: Maybe Text

If status is FAILED, Amazon Lex provides the reason that it failed to export the resource.

resourceType :: Maybe ResourceType

The type of the exported resource.

exportStatus :: Maybe ExportStatus

The status of the export.

IN_PROGRESS - The export is in progress.

READY - The export is complete.

FAILED - The export could not be completed.

url :: Maybe Text

An S3 pre-signed URL that provides the location of the exported resource. The exported resource is a ZIP archive that contains the exported resource in JSON format. The structure of the archive may change. Your code should not rely on the archive structure.

exportType :: Maybe ExportType

The format of the exported data.

name :: Maybe Text

The name of the bot being exported.

version :: Maybe Text

The version of the bot being exported.

httpStatus :: Int

The response's http status code.

newGetExportResponse

Arguments

:: Int
$sel:httpStatus:GetExportResponse'

-> GetExportResponse

Create a value of GetExportResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:GetExportResponse', getExportResponse_failureReason - If status is FAILED, Amazon Lex provides the reason that it failed to export the resource.

$sel:resourceType:GetExport', getExportResponse_resourceType - The type of the exported resource.

$sel:exportStatus:GetExportResponse', getExportResponse_exportStatus - The status of the export.

IN_PROGRESS - The export is in progress.

READY - The export is complete.

FAILED - The export could not be completed.

$sel:url:GetExportResponse', getExportResponse_url - An S3 pre-signed URL that provides the location of the exported resource. The exported resource is a ZIP archive that contains the exported resource in JSON format. The structure of the archive may change. Your code should not rely on the archive structure.

$sel:exportType:GetExport', getExportResponse_exportType - The format of the exported data.

$sel:name:GetExport', getExportResponse_name - The name of the bot being exported.

$sel:version:GetExport', getExportResponse_version - The version of the bot being exported.

$sel:httpStatus:GetExportResponse', getExportResponse_httpStatus - The response's http status code.

Response Lenses

getExportResponse_failureReason :: Lens' GetExportResponse (Maybe Text)

If status is FAILED, Amazon Lex provides the reason that it failed to export the resource.

getExportResponse_resourceType :: Lens' GetExportResponse (Maybe ResourceType)

The type of the exported resource.

getExportResponse_exportStatus :: Lens' GetExportResponse (Maybe ExportStatus)

The status of the export.

IN_PROGRESS - The export is in progress.

READY - The export is complete.

FAILED - The export could not be completed.

getExportResponse_url :: Lens' GetExportResponse (Maybe Text)

An S3 pre-signed URL that provides the location of the exported resource. The exported resource is a ZIP archive that contains the exported resource in JSON format. The structure of the archive may change. Your code should not rely on the archive structure.

getExportResponse_exportType :: Lens' GetExportResponse (Maybe ExportType)

The format of the exported data.

getExportResponse_name :: Lens' GetExportResponse (Maybe Text)

The name of the bot being exported.

getExportResponse_version :: Lens' GetExportResponse (Maybe Text)

The version of the bot being exported.

getExportResponse_httpStatus :: Lens' GetExportResponse Int

The response's http status code.