Network.AWS.LexModels.GetImport

Description

Gets information about an import job started with the StartImport operation.

Creating a Request

data GetImport

See: newGetImport smart constructor.

Constructors

GetImport'

Fields

importId :: Text

The identifier of the import job information to return.

newGetImport

Arguments

:: Text
$sel:importId:GetImport'

-> GetImport

Create a value of GetImport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:importId:GetImport', getImport_importId - The identifier of the import job information to return.

Request Lenses

getImport_importId :: Lens' GetImport Text

The identifier of the import job information to return.

Destructuring the Response

data GetImportResponse

See: newGetImportResponse smart constructor.

Constructors

GetImportResponse'

Fields

failureReason :: Maybe [Text]

A string that describes why an import job failed to complete.

resourceType :: Maybe ResourceType

The type of resource imported.

importId :: Maybe Text

The identifier for the specific import job.

createdDate :: Maybe POSIX

A timestamp for the date and time that the import job was created.

name :: Maybe Text

The name given to the import job.

mergeStrategy :: Maybe MergeStrategy

The action taken when there was a conflict between an existing resource and a resource in the import file.

importStatus :: Maybe ImportStatus

The status of the import job. If the status is FAILED, you can get the reason for the failure from the failureReason field.

httpStatus :: Int

The response's http status code.

newGetImportResponse

Arguments

:: Int
$sel:httpStatus:GetImportResponse'

-> GetImportResponse

Create a value of GetImportResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:GetImportResponse', getImportResponse_failureReason - A string that describes why an import job failed to complete.

$sel:resourceType:GetImportResponse', getImportResponse_resourceType - The type of resource imported.

$sel:importId:GetImport', getImportResponse_importId - The identifier for the specific import job.

$sel:createdDate:GetImportResponse', getImportResponse_createdDate - A timestamp for the date and time that the import job was created.

$sel:name:GetImportResponse', getImportResponse_name - The name given to the import job.

$sel:mergeStrategy:GetImportResponse', getImportResponse_mergeStrategy - The action taken when there was a conflict between an existing resource and a resource in the import file.

$sel:importStatus:GetImportResponse', getImportResponse_importStatus - The status of the import job. If the status is FAILED, you can get the reason for the failure from the failureReason field.

$sel:httpStatus:GetImportResponse', getImportResponse_httpStatus - The response's http status code.

Response Lenses

getImportResponse_failureReason :: Lens' GetImportResponse (Maybe [Text])

A string that describes why an import job failed to complete.

getImportResponse_resourceType :: Lens' GetImportResponse (Maybe ResourceType)

The type of resource imported.

getImportResponse_importId :: Lens' GetImportResponse (Maybe Text)

The identifier for the specific import job.

getImportResponse_createdDate :: Lens' GetImportResponse (Maybe UTCTime)

A timestamp for the date and time that the import job was created.

getImportResponse_name :: Lens' GetImportResponse (Maybe Text)

The name given to the import job.

getImportResponse_mergeStrategy :: Lens' GetImportResponse (Maybe MergeStrategy)

The action taken when there was a conflict between an existing resource and a resource in the import file.

getImportResponse_importStatus :: Lens' GetImportResponse (Maybe ImportStatus)

The status of the import job. If the status is FAILED, you can get the reason for the failure from the failureReason field.

getImportResponse_httpStatus :: Lens' GetImportResponse Int

The response's http status code.