Network.AWS.LexModels.GetIntent

Description

Returns information about an intent. In addition to the intent name, you must specify the intent version.

This operation requires permissions to perform the lex:GetIntent action.

Creating a Request

data GetIntent

See: newGetIntent smart constructor.

Constructors

GetIntent'

Fields

name :: Text

The name of the intent. The name is case sensitive.

version :: Text

The version of the intent.

newGetIntent

Arguments

:: Text
$sel:name:GetIntent'

-> Text
$sel:version:GetIntent'

-> GetIntent

Create a value of GetIntent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetIntent', getIntent_name - The name of the intent. The name is case sensitive.

$sel:version:GetIntent', getIntent_version - The version of the intent.

Request Lenses

getIntent_name :: Lens' GetIntent Text

The name of the intent. The name is case sensitive.

getIntent_version :: Lens' GetIntent Text

The version of the intent.

Destructuring the Response

data GetIntentResponse

See: newGetIntentResponse smart constructor.

Constructors

GetIntentResponse'

Fields

kendraConfiguration :: Maybe KendraConfiguration

Configuration information, if any, to connect to an Amazon Kendra index with the AMAZON.KendraSearchIntent intent.

createdDate :: Maybe POSIX

The date that the intent was created.

parentIntentSignature :: Maybe Text

A unique identifier for a built-in intent.

dialogCodeHook :: Maybe CodeHook

If defined in the bot, Amazon Amazon Lex invokes this Lambda function for each user input. For more information, see PutIntent.

conclusionStatement :: Maybe Statement

After the Lambda function specified in the fulfillmentActivity element fulfills the intent, Amazon Lex conveys this statement to the user.

lastUpdatedDate :: Maybe POSIX

The date that the intent was updated. When you create a resource, the creation date and the last updated date are the same.

rejectionStatement :: Maybe Statement

If the user answers "no" to the question defined in confirmationPrompt, Amazon Lex responds with this statement to acknowledge that the intent was canceled.

inputContexts :: Maybe [InputContext]

An array of InputContext objects that lists the contexts that must be active for Amazon Lex to choose the intent in a conversation with the user.

version :: Maybe Text

The version of the intent.

slots :: Maybe [Slot]

An array of intent slots configured for the intent.

fulfillmentActivity :: Maybe FulfillmentActivity

Describes how the intent is fulfilled. For more information, see PutIntent.

name :: Maybe Text

The name of the intent.

sampleUtterances :: Maybe [Text]

An array of sample utterances configured for the intent.

description :: Maybe Text

A description of the intent.

confirmationPrompt :: Maybe Prompt

If defined in the bot, Amazon Lex uses prompt to confirm the intent before fulfilling the user's request. For more information, see PutIntent.

outputContexts :: Maybe [OutputContext]

An array of OutputContext objects that lists the contexts that the intent activates when the intent is fulfilled.

followUpPrompt :: Maybe FollowUpPrompt

If defined in the bot, Amazon Lex uses this prompt to solicit additional user activity after the intent is fulfilled. For more information, see PutIntent.

checksum :: Maybe Text

Checksum of the intent.

httpStatus :: Int

The response's http status code.

newGetIntentResponse

Arguments

:: Int
$sel:httpStatus:GetIntentResponse'

-> GetIntentResponse

Create a value of GetIntentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kendraConfiguration:GetIntentResponse', getIntentResponse_kendraConfiguration - Configuration information, if any, to connect to an Amazon Kendra index with the AMAZON.KendraSearchIntent intent.

$sel:createdDate:GetIntentResponse', getIntentResponse_createdDate - The date that the intent was created.

$sel:parentIntentSignature:GetIntentResponse', getIntentResponse_parentIntentSignature - A unique identifier for a built-in intent.

$sel:dialogCodeHook:GetIntentResponse', getIntentResponse_dialogCodeHook - If defined in the bot, Amazon Amazon Lex invokes this Lambda function for each user input. For more information, see PutIntent.

$sel:conclusionStatement:GetIntentResponse', getIntentResponse_conclusionStatement - After the Lambda function specified in the fulfillmentActivity element fulfills the intent, Amazon Lex conveys this statement to the user.

$sel:lastUpdatedDate:GetIntentResponse', getIntentResponse_lastUpdatedDate - The date that the intent was updated. When you create a resource, the creation date and the last updated date are the same.

$sel:rejectionStatement:GetIntentResponse', getIntentResponse_rejectionStatement - If the user answers "no" to the question defined in confirmationPrompt, Amazon Lex responds with this statement to acknowledge that the intent was canceled.

$sel:inputContexts:GetIntentResponse', getIntentResponse_inputContexts - An array of InputContext objects that lists the contexts that must be active for Amazon Lex to choose the intent in a conversation with the user.

$sel:version:GetIntent', getIntentResponse_version - The version of the intent.

$sel:slots:GetIntentResponse', getIntentResponse_slots - An array of intent slots configured for the intent.

$sel:fulfillmentActivity:GetIntentResponse', getIntentResponse_fulfillmentActivity - Describes how the intent is fulfilled. For more information, see PutIntent.

$sel:name:GetIntent', getIntentResponse_name - The name of the intent.

$sel:sampleUtterances:GetIntentResponse', getIntentResponse_sampleUtterances - An array of sample utterances configured for the intent.

$sel:description:GetIntentResponse', getIntentResponse_description - A description of the intent.

$sel:confirmationPrompt:GetIntentResponse', getIntentResponse_confirmationPrompt - If defined in the bot, Amazon Lex uses prompt to confirm the intent before fulfilling the user's request. For more information, see PutIntent.

$sel:outputContexts:GetIntentResponse', getIntentResponse_outputContexts - An array of OutputContext objects that lists the contexts that the intent activates when the intent is fulfilled.

$sel:followUpPrompt:GetIntentResponse', getIntentResponse_followUpPrompt - If defined in the bot, Amazon Lex uses this prompt to solicit additional user activity after the intent is fulfilled. For more information, see PutIntent.

$sel:checksum:GetIntentResponse', getIntentResponse_checksum - Checksum of the intent.

$sel:httpStatus:GetIntentResponse', getIntentResponse_httpStatus - The response's http status code.

Response Lenses

getIntentResponse_kendraConfiguration :: Lens' GetIntentResponse (Maybe KendraConfiguration)

Configuration information, if any, to connect to an Amazon Kendra index with the AMAZON.KendraSearchIntent intent.

getIntentResponse_createdDate :: Lens' GetIntentResponse (Maybe UTCTime)

The date that the intent was created.

getIntentResponse_parentIntentSignature :: Lens' GetIntentResponse (Maybe Text)

A unique identifier for a built-in intent.

getIntentResponse_dialogCodeHook :: Lens' GetIntentResponse (Maybe CodeHook)

If defined in the bot, Amazon Amazon Lex invokes this Lambda function for each user input. For more information, see PutIntent.

getIntentResponse_conclusionStatement :: Lens' GetIntentResponse (Maybe Statement)

After the Lambda function specified in the fulfillmentActivity element fulfills the intent, Amazon Lex conveys this statement to the user.

getIntentResponse_lastUpdatedDate :: Lens' GetIntentResponse (Maybe UTCTime)

The date that the intent was updated. When you create a resource, the creation date and the last updated date are the same.

getIntentResponse_rejectionStatement :: Lens' GetIntentResponse (Maybe Statement)

If the user answers "no" to the question defined in confirmationPrompt, Amazon Lex responds with this statement to acknowledge that the intent was canceled.

getIntentResponse_inputContexts :: Lens' GetIntentResponse (Maybe [InputContext])

An array of InputContext objects that lists the contexts that must be active for Amazon Lex to choose the intent in a conversation with the user.

getIntentResponse_version :: Lens' GetIntentResponse (Maybe Text)

The version of the intent.

getIntentResponse_slots :: Lens' GetIntentResponse (Maybe [Slot])

An array of intent slots configured for the intent.

getIntentResponse_fulfillmentActivity :: Lens' GetIntentResponse (Maybe FulfillmentActivity)

Describes how the intent is fulfilled. For more information, see PutIntent.

getIntentResponse_name :: Lens' GetIntentResponse (Maybe Text)

The name of the intent.

getIntentResponse_sampleUtterances :: Lens' GetIntentResponse (Maybe [Text])

An array of sample utterances configured for the intent.

getIntentResponse_description :: Lens' GetIntentResponse (Maybe Text)

A description of the intent.

getIntentResponse_confirmationPrompt :: Lens' GetIntentResponse (Maybe Prompt)

If defined in the bot, Amazon Lex uses prompt to confirm the intent before fulfilling the user's request. For more information, see PutIntent.

getIntentResponse_outputContexts :: Lens' GetIntentResponse (Maybe [OutputContext])

An array of OutputContext objects that lists the contexts that the intent activates when the intent is fulfilled.

getIntentResponse_followUpPrompt :: Lens' GetIntentResponse (Maybe FollowUpPrompt)

If defined in the bot, Amazon Lex uses this prompt to solicit additional user activity after the intent is fulfilled. For more information, see PutIntent.

getIntentResponse_checksum :: Lens' GetIntentResponse (Maybe Text)

Checksum of the intent.

getIntentResponse_httpStatus :: Lens' GetIntentResponse Int

The response's http status code.