Network.AWS.LexModels.GetIntents

Description

Returns intent information as follows:

If you specify the nameContains field, returns the $LATEST version of all intents that contain the specified string.

If you don't specify the nameContains field, returns information about the $LATEST version of all intents.

The operation requires permission for the lex:GetIntents action.

This operation returns paginated results.

Creating a Request

data GetIntents

See: newGetIntents smart constructor.

Constructors

GetIntents'

Fields

nameContains :: Maybe Text

Substring to match in intent names. An intent will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

nextToken :: Maybe Text

A pagination token that fetches the next page of intents. If the response to this API call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of intents, specify the pagination token in the next request.

maxResults :: Maybe Natural

The maximum number of intents to return in the response. The default is 10.

newGetIntents :: GetIntents

Create a value of GetIntents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nameContains:GetIntents', getIntents_nameContains - Substring to match in intent names. An intent will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

$sel:nextToken:GetIntents', getIntents_nextToken - A pagination token that fetches the next page of intents. If the response to this API call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of intents, specify the pagination token in the next request.

$sel:maxResults:GetIntents', getIntents_maxResults - The maximum number of intents to return in the response. The default is 10.

Request Lenses

getIntents_nameContains :: Lens' GetIntents (Maybe Text)

Substring to match in intent names. An intent will be returned if any part of its name matches the substring. For example, "xyz" matches both "xyzabc" and "abcxyz."

getIntents_nextToken :: Lens' GetIntents (Maybe Text)

A pagination token that fetches the next page of intents. If the response to this API call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of intents, specify the pagination token in the next request.

getIntents_maxResults :: Lens' GetIntents (Maybe Natural)

The maximum number of intents to return in the response. The default is 10.

Destructuring the Response

data GetIntentsResponse

See: newGetIntentsResponse smart constructor.

Constructors

GetIntentsResponse'

Fields

nextToken :: Maybe Text

If the response is truncated, the response includes a pagination token that you can specify in your next request to fetch the next page of intents.

intents :: Maybe [IntentMetadata]

An array of Intent objects. For more information, see PutBot.

httpStatus :: Int

The response's http status code.

newGetIntentsResponse

Arguments

:: Int
$sel:httpStatus:GetIntentsResponse'

-> GetIntentsResponse

Create a value of GetIntentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetIntents', getIntentsResponse_nextToken - If the response is truncated, the response includes a pagination token that you can specify in your next request to fetch the next page of intents.

$sel:intents:GetIntentsResponse', getIntentsResponse_intents - An array of Intent objects. For more information, see PutBot.

$sel:httpStatus:GetIntentsResponse', getIntentsResponse_httpStatus - The response's http status code.

Response Lenses

getIntentsResponse_nextToken :: Lens' GetIntentsResponse (Maybe Text)

If the response is truncated, the response includes a pagination token that you can specify in your next request to fetch the next page of intents.

getIntentsResponse_intents :: Lens' GetIntentsResponse (Maybe [IntentMetadata])

An array of Intent objects. For more information, see PutBot.

getIntentsResponse_httpStatus :: Lens' GetIntentsResponse Int

The response's http status code.