Network.AWS.LexModels.GetMigrations

Description

Gets a list of migrations between Amazon Lex V1 and Amazon Lex V2.

Creating a Request

data GetMigrations

See: newGetMigrations smart constructor.

Constructors

GetMigrations'

Fields

migrationStatusEquals :: Maybe MigrationStatus

Filters the list to contain only migrations in the specified state.

nextToken :: Maybe Text

A pagination token that fetches the next page of migrations. If the response to this operation is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of migrations, specify the pagination token in the request.

maxResults :: Maybe Natural

The maximum number of migrations to return in the response. The default is 10.

sortByAttribute :: Maybe MigrationSortAttribute

The field to sort the list of migrations by. You can sort by the Amazon Lex V1 bot name or the date and time that the migration was started.

v1BotNameContains :: Maybe Text

Filters the list to contain only bots whose name contains the specified string. The string is matched anywhere in bot name.

sortByOrder :: Maybe SortOrder

The order so sort the list.

newGetMigrations :: GetMigrations

Create a value of GetMigrations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:migrationStatusEquals:GetMigrations', getMigrations_migrationStatusEquals - Filters the list to contain only migrations in the specified state.

$sel:nextToken:GetMigrations', getMigrations_nextToken - A pagination token that fetches the next page of migrations. If the response to this operation is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of migrations, specify the pagination token in the request.

$sel:maxResults:GetMigrations', getMigrations_maxResults - The maximum number of migrations to return in the response. The default is 10.

$sel:sortByAttribute:GetMigrations', getMigrations_sortByAttribute - The field to sort the list of migrations by. You can sort by the Amazon Lex V1 bot name or the date and time that the migration was started.

$sel:v1BotNameContains:GetMigrations', getMigrations_v1BotNameContains - Filters the list to contain only bots whose name contains the specified string. The string is matched anywhere in bot name.

$sel:sortByOrder:GetMigrations', getMigrations_sortByOrder - The order so sort the list.

Request Lenses

getMigrations_migrationStatusEquals :: Lens' GetMigrations (Maybe MigrationStatus)

Filters the list to contain only migrations in the specified state.

getMigrations_nextToken :: Lens' GetMigrations (Maybe Text)

A pagination token that fetches the next page of migrations. If the response to this operation is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of migrations, specify the pagination token in the request.

getMigrations_maxResults :: Lens' GetMigrations (Maybe Natural)

The maximum number of migrations to return in the response. The default is 10.

getMigrations_sortByAttribute :: Lens' GetMigrations (Maybe MigrationSortAttribute)

The field to sort the list of migrations by. You can sort by the Amazon Lex V1 bot name or the date and time that the migration was started.

getMigrations_v1BotNameContains :: Lens' GetMigrations (Maybe Text)

Filters the list to contain only bots whose name contains the specified string. The string is matched anywhere in bot name.

getMigrations_sortByOrder :: Lens' GetMigrations (Maybe SortOrder)

The order so sort the list.

Destructuring the Response

data GetMigrationsResponse

See: newGetMigrationsResponse smart constructor.

Constructors

GetMigrationsResponse'

Fields

nextToken :: Maybe Text

If the response is truncated, it includes a pagination token that you can specify in your next request to fetch the next page of migrations.

migrationSummaries :: Maybe [MigrationSummary]

An array of summaries for migrations from Amazon Lex V1 to Amazon Lex V2. To see details of the migration, use the migrationId from the summary in a call to the operation.

httpStatus :: Int

The response's http status code.

newGetMigrationsResponse

Arguments

:: Int
$sel:httpStatus:GetMigrationsResponse'

-> GetMigrationsResponse

Create a value of GetMigrationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetMigrations', getMigrationsResponse_nextToken - If the response is truncated, it includes a pagination token that you can specify in your next request to fetch the next page of migrations.

$sel:migrationSummaries:GetMigrationsResponse', getMigrationsResponse_migrationSummaries - An array of summaries for migrations from Amazon Lex V1 to Amazon Lex V2. To see details of the migration, use the migrationId from the summary in a call to the operation.

$sel:httpStatus:GetMigrationsResponse', getMigrationsResponse_httpStatus - The response's http status code.

Response Lenses

getMigrationsResponse_nextToken :: Lens' GetMigrationsResponse (Maybe Text)

If the response is truncated, it includes a pagination token that you can specify in your next request to fetch the next page of migrations.

getMigrationsResponse_migrationSummaries :: Lens' GetMigrationsResponse (Maybe [MigrationSummary])

An array of summaries for migrations from Amazon Lex V1 to Amazon Lex V2. To see details of the migration, use the migrationId from the summary in a call to the operation.

getMigrationsResponse_httpStatus :: Lens' GetMigrationsResponse Int

The response's http status code.