Network.AWS.LexModels.GetSlotType

Description

Returns information about a specific version of a slot type. In addition to specifying the slot type name, you must specify the slot type version.

This operation requires permissions for the lex:GetSlotType action.

Creating a Request

data GetSlotType

See: newGetSlotType smart constructor.

Constructors

GetSlotType'

Fields

name :: Text

The name of the slot type. The name is case sensitive.

version :: Text

The version of the slot type.

newGetSlotType

Arguments

:: Text
$sel:name:GetSlotType'

-> Text
$sel:version:GetSlotType'

-> GetSlotType

Create a value of GetSlotType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetSlotType', getSlotType_name - The name of the slot type. The name is case sensitive.

$sel:version:GetSlotType', getSlotType_version - The version of the slot type.

Request Lenses

getSlotType_name :: Lens' GetSlotType Text

The name of the slot type. The name is case sensitive.

getSlotType_version :: Lens' GetSlotType Text

The version of the slot type.

Destructuring the Response

data GetSlotTypeResponse

See: newGetSlotTypeResponse smart constructor.

Constructors

GetSlotTypeResponse'

Fields

parentSlotTypeSignature :: Maybe Text

The built-in slot type used as a parent for the slot type.

slotTypeConfigurations :: Maybe [SlotTypeConfiguration]

Configuration information that extends the parent built-in slot type.

checksum :: Maybe Text

Checksum of the $LATEST version of the slot type.

valueSelectionStrategy :: Maybe SlotValueSelectionStrategy

The strategy that Amazon Lex uses to determine the value of the slot. For more information, see PutSlotType.

createdDate :: Maybe POSIX

The date that the slot type was created.

name :: Maybe Text

The name of the slot type.

version :: Maybe Text

The version of the slot type.

lastUpdatedDate :: Maybe POSIX

The date that the slot type was updated. When you create a resource, the creation date and last update date are the same.

description :: Maybe Text

A description of the slot type.

enumerationValues :: Maybe [EnumerationValue]

A list of EnumerationValue objects that defines the values that the slot type can take.

httpStatus :: Int

The response's http status code.

newGetSlotTypeResponse

Arguments

:: Int
$sel:httpStatus:GetSlotTypeResponse'

-> GetSlotTypeResponse

Create a value of GetSlotTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parentSlotTypeSignature:GetSlotTypeResponse', getSlotTypeResponse_parentSlotTypeSignature - The built-in slot type used as a parent for the slot type.

$sel:slotTypeConfigurations:GetSlotTypeResponse', getSlotTypeResponse_slotTypeConfigurations - Configuration information that extends the parent built-in slot type.

$sel:checksum:GetSlotTypeResponse', getSlotTypeResponse_checksum - Checksum of the $LATEST version of the slot type.

$sel:valueSelectionStrategy:GetSlotTypeResponse', getSlotTypeResponse_valueSelectionStrategy - The strategy that Amazon Lex uses to determine the value of the slot. For more information, see PutSlotType.

$sel:createdDate:GetSlotTypeResponse', getSlotTypeResponse_createdDate - The date that the slot type was created.

$sel:name:GetSlotType', getSlotTypeResponse_name - The name of the slot type.

$sel:version:GetSlotType', getSlotTypeResponse_version - The version of the slot type.

$sel:lastUpdatedDate:GetSlotTypeResponse', getSlotTypeResponse_lastUpdatedDate - The date that the slot type was updated. When you create a resource, the creation date and last update date are the same.

$sel:description:GetSlotTypeResponse', getSlotTypeResponse_description - A description of the slot type.

$sel:enumerationValues:GetSlotTypeResponse', getSlotTypeResponse_enumerationValues - A list of EnumerationValue objects that defines the values that the slot type can take.

$sel:httpStatus:GetSlotTypeResponse', getSlotTypeResponse_httpStatus - The response's http status code.

Response Lenses

getSlotTypeResponse_parentSlotTypeSignature :: Lens' GetSlotTypeResponse (Maybe Text)

The built-in slot type used as a parent for the slot type.

getSlotTypeResponse_slotTypeConfigurations :: Lens' GetSlotTypeResponse (Maybe [SlotTypeConfiguration])

Configuration information that extends the parent built-in slot type.

getSlotTypeResponse_checksum :: Lens' GetSlotTypeResponse (Maybe Text)

Checksum of the $LATEST version of the slot type.

getSlotTypeResponse_valueSelectionStrategy :: Lens' GetSlotTypeResponse (Maybe SlotValueSelectionStrategy)

The strategy that Amazon Lex uses to determine the value of the slot. For more information, see PutSlotType.

getSlotTypeResponse_createdDate :: Lens' GetSlotTypeResponse (Maybe UTCTime)

The date that the slot type was created.

getSlotTypeResponse_name :: Lens' GetSlotTypeResponse (Maybe Text)

The name of the slot type.

getSlotTypeResponse_version :: Lens' GetSlotTypeResponse (Maybe Text)

The version of the slot type.

getSlotTypeResponse_lastUpdatedDate :: Lens' GetSlotTypeResponse (Maybe UTCTime)

The date that the slot type was updated. When you create a resource, the creation date and last update date are the same.

getSlotTypeResponse_description :: Lens' GetSlotTypeResponse (Maybe Text)

A description of the slot type.

getSlotTypeResponse_enumerationValues :: Lens' GetSlotTypeResponse (Maybe [EnumerationValue])

A list of EnumerationValue objects that defines the values that the slot type can take.

getSlotTypeResponse_httpStatus :: Lens' GetSlotTypeResponse Int

The response's http status code.