Network.AWS.LexModels.GetSlotTypeVersions

Description

Gets information about all versions of a slot type.

The GetSlotTypeVersions operation returns a SlotTypeMetadata object for each version of a slot type. For example, if a slot type has three numbered versions, the GetSlotTypeVersions operation returns four SlotTypeMetadata objects in the response, one for each numbered version and one for the $LATEST version.

The GetSlotTypeVersions operation always returns at least one version, the $LATEST version.

This operation requires permissions for the lex:GetSlotTypeVersions action.

This operation returns paginated results.

Creating a Request

data GetSlotTypeVersions

See: newGetSlotTypeVersions smart constructor.

Constructors

GetSlotTypeVersions'

Fields

nextToken :: Maybe Text

A pagination token for fetching the next page of slot type versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

maxResults :: Maybe Natural

The maximum number of slot type versions to return in the response. The default is 10.

name :: Text

The name of the slot type for which versions should be returned.

newGetSlotTypeVersions

Arguments

:: Text
$sel:name:GetSlotTypeVersions'

-> GetSlotTypeVersions

Create a value of GetSlotTypeVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetSlotTypeVersions', getSlotTypeVersions_nextToken - A pagination token for fetching the next page of slot type versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

$sel:maxResults:GetSlotTypeVersions', getSlotTypeVersions_maxResults - The maximum number of slot type versions to return in the response. The default is 10.

$sel:name:GetSlotTypeVersions', getSlotTypeVersions_name - The name of the slot type for which versions should be returned.

Request Lenses

getSlotTypeVersions_nextToken :: Lens' GetSlotTypeVersions (Maybe Text)

A pagination token for fetching the next page of slot type versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

getSlotTypeVersions_maxResults :: Lens' GetSlotTypeVersions (Maybe Natural)

The maximum number of slot type versions to return in the response. The default is 10.

getSlotTypeVersions_name :: Lens' GetSlotTypeVersions Text

The name of the slot type for which versions should be returned.

Destructuring the Response

data GetSlotTypeVersionsResponse

See: newGetSlotTypeVersionsResponse smart constructor.

Constructors

GetSlotTypeVersionsResponse'

Fields

slotTypes :: Maybe [SlotTypeMetadata]

An array of SlotTypeMetadata objects, one for each numbered version of the slot type plus one for the $LATEST version.

nextToken :: Maybe Text

A pagination token for fetching the next page of slot type versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

httpStatus :: Int

The response's http status code.

newGetSlotTypeVersionsResponse

Arguments

:: Int
$sel:httpStatus:GetSlotTypeVersionsResponse'

-> GetSlotTypeVersionsResponse

Create a value of GetSlotTypeVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:slotTypes:GetSlotTypeVersionsResponse', getSlotTypeVersionsResponse_slotTypes - An array of SlotTypeMetadata objects, one for each numbered version of the slot type plus one for the $LATEST version.

$sel:nextToken:GetSlotTypeVersions', getSlotTypeVersionsResponse_nextToken - A pagination token for fetching the next page of slot type versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

$sel:httpStatus:GetSlotTypeVersionsResponse', getSlotTypeVersionsResponse_httpStatus - The response's http status code.

Response Lenses

getSlotTypeVersionsResponse_slotTypes :: Lens' GetSlotTypeVersionsResponse (Maybe [SlotTypeMetadata])

An array of SlotTypeMetadata objects, one for each numbered version of the slot type plus one for the $LATEST version.

getSlotTypeVersionsResponse_nextToken :: Lens' GetSlotTypeVersionsResponse (Maybe Text)

A pagination token for fetching the next page of slot type versions. If the response to this call is truncated, Amazon Lex returns a pagination token in the response. To fetch the next page of versions, specify the pagination token in the next request.

getSlotTypeVersionsResponse_httpStatus :: Lens' GetSlotTypeVersionsResponse Int

The response's http status code.