Network.AWS.LexModels.PutBot

Description

Creates an Amazon Lex conversational bot or replaces an existing bot. When you create or update a bot you are only required to specify a name, a locale, and whether the bot is directed toward children under age 13. You can use this to add intents later, or to remove intents from an existing bot. When you create a bot with the minimum information, the bot is created or updated but Amazon Lex returns the @ response FAILED@. You can build the bot after you add one or more intents. For more information about Amazon Lex bots, see how-it-works.

If you specify the name of an existing bot, the fields in the request replace the existing values in the $LATEST version of the bot. Amazon Lex removes any fields that you don't provide values for in the request, except for the idleTTLInSeconds and privacySettings fields, which are set to their default values. If you don't specify values for required fields, Amazon Lex throws an exception.

This operation requires permissions for the lex:PutBot action. For more information, see security-iam.

Creating a Request

data PutBot

See: newPutBot smart constructor.

Constructors

PutBot'

Fields

abortStatement :: Maybe Statement

When Amazon Lex can't understand the user's input in context, it tries to elicit the information a few times. After that, Amazon Lex sends the message defined in abortStatement to the user, and then cancels the conversation. To set the number of retries, use the valueElicitationPrompt field for the slot type.

For example, in a pizza ordering bot, Amazon Lex might ask a user "What type of crust would you like?" If the user's response is not one of the expected responses (for example, "thin crust, "deep dish," etc.), Amazon Lex tries to elicit a correct response a few more times.

For example, in a pizza ordering application, OrderPizza might be one of the intents. This intent might require the CrustType slot. You specify the valueElicitationPrompt field when you create the CrustType slot.

If you have defined a fallback intent the cancel statement will not be sent to the user, the fallback intent is used instead. For more information, see AMAZON.FallbackIntent.

processBehavior :: Maybe ProcessBehavior

If you set the processBehavior element to BUILD, Amazon Lex builds the bot so that it can be run. If you set the element to SAVE Amazon Lex saves the bot, but doesn't build it.

If you don't specify this value, the default value is BUILD.

voiceId :: Maybe Text

The Amazon Polly voice ID that you want Amazon Lex to use for voice interactions with the user. The locale configured for the voice must match the locale of the bot. For more information, see Voices in Amazon Polly in the Amazon Polly Developer Guide.

clarificationPrompt :: Maybe Prompt

When Amazon Lex doesn't understand the user's intent, it uses this message to get clarification. To specify how many times Amazon Lex should repeat the clarification prompt, use the maxAttempts field. If Amazon Lex still doesn't understand, it sends the message in the abortStatement field.

When you create a clarification prompt, make sure that it suggests the correct response from the user. for example, for a bot that orders pizza and drinks, you might create this clarification prompt: "What would you like to do? You can say 'Order a pizza' or 'Order a drink.'"

If you have defined a fallback intent, it will be invoked if the clarification prompt is repeated the number of times defined in the maxAttempts field. For more information, see AMAZON.FallbackIntent.

If you don't define a clarification prompt, at runtime Amazon Lex will return a 400 Bad Request exception in three cases:

Follow-up prompt - When the user responds to a follow-up prompt but does not provide an intent. For example, in response to a follow-up prompt that says "Would you like anything else today?" the user says "Yes." Amazon Lex will return a 400 Bad Request exception because it does not have a clarification prompt to send to the user to get an intent.

Lambda function - When using a Lambda function, you return an ElicitIntent dialog type. Since Amazon Lex does not have a clarification prompt to get an intent from the user, it returns a 400 Bad Request exception.

PutSession operation - When using the PutSession operation, you send an ElicitIntent dialog type. Since Amazon Lex does not have a clarification prompt to get an intent from the user, it returns a 400 Bad Request exception.

nluIntentConfidenceThreshold :: Maybe Double

Determines the threshold where Amazon Lex will insert the AMAZON.FallbackIntent, AMAZON.KendraSearchIntent, or both when returning alternative intents in a PostContent or PostText response. AMAZON.FallbackIntent and AMAZON.KendraSearchIntent are only inserted if they are configured for the bot.

You must set the enableModelImprovements parameter to true to use confidence scores in the following regions.

US East (N. Virginia) (us-east-1)

US West (Oregon) (us-west-2)

Asia Pacific (Sydney) (ap-southeast-2)

EU (Ireland) (eu-west-1)

In other Regions, the enableModelImprovements parameter is set to true by default.

For example, suppose a bot is configured with the confidence threshold of 0.80 and the AMAZON.FallbackIntent. Amazon Lex returns three alternative intents with the following confidence scores: IntentA (0.70), IntentB (0.60), IntentC (0.50). The response from the PostText operation would be:

AMAZON.FallbackIntent

IntentA

IntentB

IntentC

enableModelImprovements :: Maybe Bool

Set to true to enable access to natural language understanding improvements.

When you set the enableModelImprovements parameter to true you can use the nluIntentConfidenceThreshold parameter to configure confidence scores. For more information, see Confidence Scores.

You can only set the enableModelImprovements parameter in certain Regions. If you set the parameter to true, your bot has access to accuracy improvements.

The Regions where you can set the enableModelImprovements parameter to true are:

US East (N. Virginia) (us-east-1)

US West (Oregon) (us-west-2)

Asia Pacific (Sydney) (ap-southeast-2)

EU (Ireland) (eu-west-1)

In other Regions, the enableModelImprovements parameter is set to true by default. In these Regions setting the parameter to false throws a ValidationException exception.

idleSessionTTLInSeconds :: Maybe Natural

The maximum time in seconds that Amazon Lex retains the data gathered in a conversation.

A user interaction session remains active for the amount of time specified. If no conversation occurs during this time, the session expires and Amazon Lex deletes any data provided before the timeout.

For example, suppose that a user chooses the OrderPizza intent, but gets sidetracked halfway through placing an order. If the user doesn't complete the order within the specified time, Amazon Lex discards the slot information that it gathered, and the user must start over.

If you don't include the idleSessionTTLInSeconds element in a PutBot operation request, Amazon Lex uses the default value. This is also true if the request replaces an existing bot.

The default is 300 seconds (5 minutes).

intents :: Maybe [Intent]

An array of Intent objects. Each intent represents a command that a user can express. For example, a pizza ordering bot might support an OrderPizza intent. For more information, see how-it-works.

tags :: Maybe [Tag]

A list of tags to add to the bot. You can only add tags when you create a bot, you can't use the PutBot operation to update the tags on a bot. To update tags, use the TagResource operation.

description :: Maybe Text

A description of the bot.

createVersion :: Maybe Bool

When set to true a new numbered version of the bot is created. This is the same as calling the CreateBotVersion operation. If you don't specify createVersion, the default is false.

detectSentiment :: Maybe Bool

When set to true user utterances are sent to Amazon Comprehend for sentiment analysis. If you don't specify detectSentiment, the default is false.

checksum :: Maybe Text

Identifies a specific revision of the $LATEST version.

When you create a new bot, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception.

When you want to update a bot, set the checksum field to the checksum of the most recent revision of the $LATEST version. If you don't specify the checksum field, or if the checksum does not match the $LATEST version, you get a PreconditionFailedException exception.

name :: Text

The name of the bot. The name is not case sensitive.

locale :: Locale

Specifies the target locale for the bot. Any intent used in the bot must be compatible with the locale of the bot.

The default is en-US.

childDirected :: Bool

For each Amazon Lex bot created with the Amazon Lex Model Building Service, you must specify whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to the Children's Online Privacy Protection Act (COPPA) by specifying true or false in the childDirected field. By specifying true in the childDirected field, you confirm that your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. By specifying false in the childDirected field, you confirm that your use of Amazon Lex is not related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. You may not specify a default value for the childDirected field that does not accurately reflect whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA.

If your use of Amazon Lex relates to a website, program, or other application that is directed in whole or in part, to children under age 13, you must obtain any required verifiable parental consent under COPPA. For information regarding the use of Amazon Lex in connection with websites, programs, or other applications that are directed or targeted, in whole or in part, to children under age 13, see the Amazon Lex FAQ.

newPutBot

Arguments

:: Text
$sel:name:PutBot'

-> Locale
$sel:locale:PutBot'

-> Bool
$sel:childDirected:PutBot'

-> PutBot

Create a value of PutBot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:abortStatement:PutBot', putBot_abortStatement - When Amazon Lex can't understand the user's input in context, it tries to elicit the information a few times. After that, Amazon Lex sends the message defined in abortStatement to the user, and then cancels the conversation. To set the number of retries, use the valueElicitationPrompt field for the slot type.

For example, in a pizza ordering bot, Amazon Lex might ask a user "What type of crust would you like?" If the user's response is not one of the expected responses (for example, "thin crust, "deep dish," etc.), Amazon Lex tries to elicit a correct response a few more times.

For example, in a pizza ordering application, OrderPizza might be one of the intents. This intent might require the CrustType slot. You specify the valueElicitationPrompt field when you create the CrustType slot.

If you have defined a fallback intent the cancel statement will not be sent to the user, the fallback intent is used instead. For more information, see AMAZON.FallbackIntent.

$sel:processBehavior:PutBot', putBot_processBehavior - If you set the processBehavior element to BUILD, Amazon Lex builds the bot so that it can be run. If you set the element to SAVE Amazon Lex saves the bot, but doesn't build it.

If you don't specify this value, the default value is BUILD.

$sel:voiceId:PutBot', putBot_voiceId - The Amazon Polly voice ID that you want Amazon Lex to use for voice interactions with the user. The locale configured for the voice must match the locale of the bot. For more information, see Voices in Amazon Polly in the Amazon Polly Developer Guide.

$sel:clarificationPrompt:PutBot', putBot_clarificationPrompt - When Amazon Lex doesn't understand the user's intent, it uses this message to get clarification. To specify how many times Amazon Lex should repeat the clarification prompt, use the maxAttempts field. If Amazon Lex still doesn't understand, it sends the message in the abortStatement field.

When you create a clarification prompt, make sure that it suggests the correct response from the user. for example, for a bot that orders pizza and drinks, you might create this clarification prompt: "What would you like to do? You can say 'Order a pizza' or 'Order a drink.'"

If you have defined a fallback intent, it will be invoked if the clarification prompt is repeated the number of times defined in the maxAttempts field. For more information, see AMAZON.FallbackIntent.

If you don't define a clarification prompt, at runtime Amazon Lex will return a 400 Bad Request exception in three cases:

Follow-up prompt - When the user responds to a follow-up prompt but does not provide an intent. For example, in response to a follow-up prompt that says "Would you like anything else today?" the user says "Yes." Amazon Lex will return a 400 Bad Request exception because it does not have a clarification prompt to send to the user to get an intent.

Lambda function - When using a Lambda function, you return an ElicitIntent dialog type. Since Amazon Lex does not have a clarification prompt to get an intent from the user, it returns a 400 Bad Request exception.

PutSession operation - When using the PutSession operation, you send an ElicitIntent dialog type. Since Amazon Lex does not have a clarification prompt to get an intent from the user, it returns a 400 Bad Request exception.

$sel:nluIntentConfidenceThreshold:PutBot', putBot_nluIntentConfidenceThreshold - Determines the threshold where Amazon Lex will insert the AMAZON.FallbackIntent, AMAZON.KendraSearchIntent, or both when returning alternative intents in a PostContent or PostText response. AMAZON.FallbackIntent and AMAZON.KendraSearchIntent are only inserted if they are configured for the bot.

You must set the enableModelImprovements parameter to true to use confidence scores in the following regions.

US East (N. Virginia) (us-east-1)

US West (Oregon) (us-west-2)

Asia Pacific (Sydney) (ap-southeast-2)

EU (Ireland) (eu-west-1)

In other Regions, the enableModelImprovements parameter is set to true by default.

For example, suppose a bot is configured with the confidence threshold of 0.80 and the AMAZON.FallbackIntent. Amazon Lex returns three alternative intents with the following confidence scores: IntentA (0.70), IntentB (0.60), IntentC (0.50). The response from the PostText operation would be:

AMAZON.FallbackIntent

IntentA

IntentB

IntentC

$sel:enableModelImprovements:PutBot', putBot_enableModelImprovements - Set to true to enable access to natural language understanding improvements.

When you set the enableModelImprovements parameter to true you can use the nluIntentConfidenceThreshold parameter to configure confidence scores. For more information, see Confidence Scores.

You can only set the enableModelImprovements parameter in certain Regions. If you set the parameter to true, your bot has access to accuracy improvements.

The Regions where you can set the enableModelImprovements parameter to true are:

US East (N. Virginia) (us-east-1)

US West (Oregon) (us-west-2)

Asia Pacific (Sydney) (ap-southeast-2)

EU (Ireland) (eu-west-1)

In other Regions, the enableModelImprovements parameter is set to true by default. In these Regions setting the parameter to false throws a ValidationException exception.

$sel:idleSessionTTLInSeconds:PutBot', putBot_idleSessionTTLInSeconds - The maximum time in seconds that Amazon Lex retains the data gathered in a conversation.

A user interaction session remains active for the amount of time specified. If no conversation occurs during this time, the session expires and Amazon Lex deletes any data provided before the timeout.

For example, suppose that a user chooses the OrderPizza intent, but gets sidetracked halfway through placing an order. If the user doesn't complete the order within the specified time, Amazon Lex discards the slot information that it gathered, and the user must start over.

If you don't include the idleSessionTTLInSeconds element in a PutBot operation request, Amazon Lex uses the default value. This is also true if the request replaces an existing bot.

The default is 300 seconds (5 minutes).

$sel:intents:PutBot', putBot_intents - An array of Intent objects. Each intent represents a command that a user can express. For example, a pizza ordering bot might support an OrderPizza intent. For more information, see how-it-works.

$sel:tags:PutBot', putBot_tags - A list of tags to add to the bot. You can only add tags when you create a bot, you can't use the PutBot operation to update the tags on a bot. To update tags, use the TagResource operation.

$sel:description:PutBot', putBot_description - A description of the bot.

$sel:createVersion:PutBot', putBot_createVersion - When set to true a new numbered version of the bot is created. This is the same as calling the CreateBotVersion operation. If you don't specify createVersion, the default is false.

$sel:detectSentiment:PutBot', putBot_detectSentiment - When set to true user utterances are sent to Amazon Comprehend for sentiment analysis. If you don't specify detectSentiment, the default is false.

$sel:checksum:PutBot', putBot_checksum - Identifies a specific revision of the $LATEST version.

When you create a new bot, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception.

When you want to update a bot, set the checksum field to the checksum of the most recent revision of the $LATEST version. If you don't specify the checksum field, or if the checksum does not match the $LATEST version, you get a PreconditionFailedException exception.

$sel:name:PutBot', putBot_name - The name of the bot. The name is not case sensitive.

$sel:locale:PutBot', putBot_locale - Specifies the target locale for the bot. Any intent used in the bot must be compatible with the locale of the bot.

The default is en-US.

$sel:childDirected:PutBot', putBot_childDirected - For each Amazon Lex bot created with the Amazon Lex Model Building Service, you must specify whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to the Children's Online Privacy Protection Act (COPPA) by specifying true or false in the childDirected field. By specifying true in the childDirected field, you confirm that your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. By specifying false in the childDirected field, you confirm that your use of Amazon Lex is not related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. You may not specify a default value for the childDirected field that does not accurately reflect whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA.

If your use of Amazon Lex relates to a website, program, or other application that is directed in whole or in part, to children under age 13, you must obtain any required verifiable parental consent under COPPA. For information regarding the use of Amazon Lex in connection with websites, programs, or other applications that are directed or targeted, in whole or in part, to children under age 13, see the Amazon Lex FAQ.

Request Lenses

putBot_abortStatement :: Lens' PutBot (Maybe Statement)

When Amazon Lex can't understand the user's input in context, it tries to elicit the information a few times. After that, Amazon Lex sends the message defined in abortStatement to the user, and then cancels the conversation. To set the number of retries, use the valueElicitationPrompt field for the slot type.

For example, in a pizza ordering bot, Amazon Lex might ask a user "What type of crust would you like?" If the user's response is not one of the expected responses (for example, "thin crust, "deep dish," etc.), Amazon Lex tries to elicit a correct response a few more times.

For example, in a pizza ordering application, OrderPizza might be one of the intents. This intent might require the CrustType slot. You specify the valueElicitationPrompt field when you create the CrustType slot.

If you have defined a fallback intent the cancel statement will not be sent to the user, the fallback intent is used instead. For more information, see AMAZON.FallbackIntent.

putBot_processBehavior :: Lens' PutBot (Maybe ProcessBehavior)

If you set the processBehavior element to BUILD, Amazon Lex builds the bot so that it can be run. If you set the element to SAVE Amazon Lex saves the bot, but doesn't build it.

If you don't specify this value, the default value is BUILD.

putBot_voiceId :: Lens' PutBot (Maybe Text)

The Amazon Polly voice ID that you want Amazon Lex to use for voice interactions with the user. The locale configured for the voice must match the locale of the bot. For more information, see Voices in Amazon Polly in the Amazon Polly Developer Guide.

putBot_clarificationPrompt :: Lens' PutBot (Maybe Prompt)

When Amazon Lex doesn't understand the user's intent, it uses this message to get clarification. To specify how many times Amazon Lex should repeat the clarification prompt, use the maxAttempts field. If Amazon Lex still doesn't understand, it sends the message in the abortStatement field.

When you create a clarification prompt, make sure that it suggests the correct response from the user. for example, for a bot that orders pizza and drinks, you might create this clarification prompt: "What would you like to do? You can say 'Order a pizza' or 'Order a drink.'"

If you have defined a fallback intent, it will be invoked if the clarification prompt is repeated the number of times defined in the maxAttempts field. For more information, see AMAZON.FallbackIntent.

If you don't define a clarification prompt, at runtime Amazon Lex will return a 400 Bad Request exception in three cases:

Follow-up prompt - When the user responds to a follow-up prompt but does not provide an intent. For example, in response to a follow-up prompt that says "Would you like anything else today?" the user says "Yes." Amazon Lex will return a 400 Bad Request exception because it does not have a clarification prompt to send to the user to get an intent.

Lambda function - When using a Lambda function, you return an ElicitIntent dialog type. Since Amazon Lex does not have a clarification prompt to get an intent from the user, it returns a 400 Bad Request exception.

PutSession operation - When using the PutSession operation, you send an ElicitIntent dialog type. Since Amazon Lex does not have a clarification prompt to get an intent from the user, it returns a 400 Bad Request exception.

putBot_nluIntentConfidenceThreshold :: Lens' PutBot (Maybe Double)

Determines the threshold where Amazon Lex will insert the AMAZON.FallbackIntent, AMAZON.KendraSearchIntent, or both when returning alternative intents in a PostContent or PostText response. AMAZON.FallbackIntent and AMAZON.KendraSearchIntent are only inserted if they are configured for the bot.

You must set the enableModelImprovements parameter to true to use confidence scores in the following regions.

US East (N. Virginia) (us-east-1)

US West (Oregon) (us-west-2)

Asia Pacific (Sydney) (ap-southeast-2)

EU (Ireland) (eu-west-1)

In other Regions, the enableModelImprovements parameter is set to true by default.

For example, suppose a bot is configured with the confidence threshold of 0.80 and the AMAZON.FallbackIntent. Amazon Lex returns three alternative intents with the following confidence scores: IntentA (0.70), IntentB (0.60), IntentC (0.50). The response from the PostText operation would be:

AMAZON.FallbackIntent

IntentA

IntentB

IntentC

putBot_enableModelImprovements :: Lens' PutBot (Maybe Bool)

Set to true to enable access to natural language understanding improvements.

When you set the enableModelImprovements parameter to true you can use the nluIntentConfidenceThreshold parameter to configure confidence scores. For more information, see Confidence Scores.

You can only set the enableModelImprovements parameter in certain Regions. If you set the parameter to true, your bot has access to accuracy improvements.

The Regions where you can set the enableModelImprovements parameter to true are:

US East (N. Virginia) (us-east-1)

US West (Oregon) (us-west-2)

Asia Pacific (Sydney) (ap-southeast-2)

EU (Ireland) (eu-west-1)

In other Regions, the enableModelImprovements parameter is set to true by default. In these Regions setting the parameter to false throws a ValidationException exception.

putBot_idleSessionTTLInSeconds :: Lens' PutBot (Maybe Natural)

The maximum time in seconds that Amazon Lex retains the data gathered in a conversation.

A user interaction session remains active for the amount of time specified. If no conversation occurs during this time, the session expires and Amazon Lex deletes any data provided before the timeout.

For example, suppose that a user chooses the OrderPizza intent, but gets sidetracked halfway through placing an order. If the user doesn't complete the order within the specified time, Amazon Lex discards the slot information that it gathered, and the user must start over.

If you don't include the idleSessionTTLInSeconds element in a PutBot operation request, Amazon Lex uses the default value. This is also true if the request replaces an existing bot.

The default is 300 seconds (5 minutes).

putBot_intents :: Lens' PutBot (Maybe [Intent])

An array of Intent objects. Each intent represents a command that a user can express. For example, a pizza ordering bot might support an OrderPizza intent. For more information, see how-it-works.

putBot_tags :: Lens' PutBot (Maybe [Tag])

A list of tags to add to the bot. You can only add tags when you create a bot, you can't use the PutBot operation to update the tags on a bot. To update tags, use the TagResource operation.

putBot_description :: Lens' PutBot (Maybe Text)

A description of the bot.

putBot_createVersion :: Lens' PutBot (Maybe Bool)

When set to true a new numbered version of the bot is created. This is the same as calling the CreateBotVersion operation. If you don't specify createVersion, the default is false.

putBot_detectSentiment :: Lens' PutBot (Maybe Bool)

When set to true user utterances are sent to Amazon Comprehend for sentiment analysis. If you don't specify detectSentiment, the default is false.

putBot_checksum :: Lens' PutBot (Maybe Text)

Identifies a specific revision of the $LATEST version.

When you create a new bot, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception.

When you want to update a bot, set the checksum field to the checksum of the most recent revision of the $LATEST version. If you don't specify the checksum field, or if the checksum does not match the $LATEST version, you get a PreconditionFailedException exception.

putBot_name :: Lens' PutBot Text

The name of the bot. The name is not case sensitive.

putBot_locale :: Lens' PutBot Locale

Specifies the target locale for the bot. Any intent used in the bot must be compatible with the locale of the bot.

The default is en-US.

putBot_childDirected :: Lens' PutBot Bool

For each Amazon Lex bot created with the Amazon Lex Model Building Service, you must specify whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to the Children's Online Privacy Protection Act (COPPA) by specifying true or false in the childDirected field. By specifying true in the childDirected field, you confirm that your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. By specifying false in the childDirected field, you confirm that your use of Amazon Lex is not related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. You may not specify a default value for the childDirected field that does not accurately reflect whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA.

If your use of Amazon Lex relates to a website, program, or other application that is directed in whole or in part, to children under age 13, you must obtain any required verifiable parental consent under COPPA. For information regarding the use of Amazon Lex in connection with websites, programs, or other applications that are directed or targeted, in whole or in part, to children under age 13, see the Amazon Lex FAQ.

Destructuring the Response

data PutBotResponse

See: newPutBotResponse smart constructor.

Constructors

PutBotResponse'

Fields

abortStatement :: Maybe Statement

The message that Amazon Lex uses to cancel a conversation. For more information, see PutBot.

createdDate :: Maybe POSIX

The date that the bot was created.

status :: Maybe LexStatus

When you send a request to create a bot with processBehavior set to BUILD, Amazon Lex sets the status response element to BUILDING.

In the READY_BASIC_TESTING state you can test the bot with user inputs that exactly match the utterances configured for the bot's intents and values in the slot types.

If Amazon Lex can't build the bot, Amazon Lex sets status to FAILED. Amazon Lex returns the reason for the failure in the failureReason response element.

When you set processBehavior to SAVE, Amazon Lex sets the status code to NOT BUILT.

When the bot is in the READY state you can test and publish the bot.

lastUpdatedDate :: Maybe POSIX

The date that the bot was updated. When you create a resource, the creation date and last updated date are the same.

voiceId :: Maybe Text

The Amazon Polly voice ID that Amazon Lex uses for voice interaction with the user. For more information, see PutBot.

clarificationPrompt :: Maybe Prompt

The prompts that Amazon Lex uses when it doesn't understand the user's intent. For more information, see PutBot.

nluIntentConfidenceThreshold :: Maybe Double

The score that determines where Amazon Lex inserts the AMAZON.FallbackIntent, AMAZON.KendraSearchIntent, or both when returning alternative intents in a PostContent or PostText response. AMAZON.FallbackIntent is inserted if the confidence score for all intents is below this value. AMAZON.KendraSearchIntent is only inserted if it is configured for the bot.

locale :: Maybe Locale

The target locale for the bot.

enableModelImprovements :: Maybe Bool

Indicates whether the bot uses accuracy improvements. true indicates that the bot is using the improvements, otherwise, false.

version :: Maybe Text

The version of the bot. For a new bot, the version is always $LATEST.

idleSessionTTLInSeconds :: Maybe Natural

The maximum length of time that Amazon Lex retains the data gathered in a conversation. For more information, see PutBot.

name :: Maybe Text

The name of the bot.

intents :: Maybe [Intent]

An array of Intent objects. For more information, see PutBot.

failureReason :: Maybe Text

If status is FAILED, Amazon Lex provides the reason that it failed to build the bot.

tags :: Maybe [Tag]

A list of tags associated with the bot.

childDirected :: Maybe Bool

For each Amazon Lex bot created with the Amazon Lex Model Building Service, you must specify whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to the Children's Online Privacy Protection Act (COPPA) by specifying true or false in the childDirected field. By specifying true in the childDirected field, you confirm that your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. By specifying false in the childDirected field, you confirm that your use of Amazon Lex is not related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. You may not specify a default value for the childDirected field that does not accurately reflect whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA.

If your use of Amazon Lex relates to a website, program, or other application that is directed in whole or in part, to children under age 13, you must obtain any required verifiable parental consent under COPPA. For information regarding the use of Amazon Lex in connection with websites, programs, or other applications that are directed or targeted, in whole or in part, to children under age 13, see the Amazon Lex FAQ.

description :: Maybe Text

A description of the bot.

createVersion :: Maybe Bool

True if a new version of the bot was created. If the createVersion field was not specified in the request, the createVersion field is set to false in the response.

detectSentiment :: Maybe Bool

true if the bot is configured to send user utterances to Amazon Comprehend for sentiment analysis. If the detectSentiment field was not specified in the request, the detectSentiment field is false in the response.

checksum :: Maybe Text

Checksum of the bot that you created.

httpStatus :: Int

The response's http status code.

newPutBotResponse

Arguments

:: Int
$sel:httpStatus:PutBotResponse'

-> PutBotResponse

Create a value of PutBotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:abortStatement:PutBot', putBotResponse_abortStatement - The message that Amazon Lex uses to cancel a conversation. For more information, see PutBot.

$sel:createdDate:PutBotResponse', putBotResponse_createdDate - The date that the bot was created.

$sel:status:PutBotResponse', putBotResponse_status - When you send a request to create a bot with processBehavior set to BUILD, Amazon Lex sets the status response element to BUILDING.

In the READY_BASIC_TESTING state you can test the bot with user inputs that exactly match the utterances configured for the bot's intents and values in the slot types.

If Amazon Lex can't build the bot, Amazon Lex sets status to FAILED. Amazon Lex returns the reason for the failure in the failureReason response element.

When you set processBehavior to SAVE, Amazon Lex sets the status code to NOT BUILT.

When the bot is in the READY state you can test and publish the bot.

$sel:lastUpdatedDate:PutBotResponse', putBotResponse_lastUpdatedDate - The date that the bot was updated. When you create a resource, the creation date and last updated date are the same.

$sel:voiceId:PutBot', putBotResponse_voiceId - The Amazon Polly voice ID that Amazon Lex uses for voice interaction with the user. For more information, see PutBot.

$sel:clarificationPrompt:PutBot', putBotResponse_clarificationPrompt - The prompts that Amazon Lex uses when it doesn't understand the user's intent. For more information, see PutBot.

$sel:nluIntentConfidenceThreshold:PutBot', putBotResponse_nluIntentConfidenceThreshold - The score that determines where Amazon Lex inserts the AMAZON.FallbackIntent, AMAZON.KendraSearchIntent, or both when returning alternative intents in a PostContent or PostText response. AMAZON.FallbackIntent is inserted if the confidence score for all intents is below this value. AMAZON.KendraSearchIntent is only inserted if it is configured for the bot.

$sel:locale:PutBot', putBotResponse_locale - The target locale for the bot.

$sel:enableModelImprovements:PutBot', putBotResponse_enableModelImprovements - Indicates whether the bot uses accuracy improvements. true indicates that the bot is using the improvements, otherwise, false.

$sel:version:PutBotResponse', putBotResponse_version - The version of the bot. For a new bot, the version is always $LATEST.

$sel:idleSessionTTLInSeconds:PutBot', putBotResponse_idleSessionTTLInSeconds - The maximum length of time that Amazon Lex retains the data gathered in a conversation. For more information, see PutBot.

$sel:name:PutBot', putBotResponse_name - The name of the bot.

$sel:intents:PutBot', putBotResponse_intents - An array of Intent objects. For more information, see PutBot.

$sel:failureReason:PutBotResponse', putBotResponse_failureReason - If status is FAILED, Amazon Lex provides the reason that it failed to build the bot.

$sel:tags:PutBot', putBotResponse_tags - A list of tags associated with the bot.

$sel:childDirected:PutBot', putBotResponse_childDirected - For each Amazon Lex bot created with the Amazon Lex Model Building Service, you must specify whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to the Children's Online Privacy Protection Act (COPPA) by specifying true or false in the childDirected field. By specifying true in the childDirected field, you confirm that your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. By specifying false in the childDirected field, you confirm that your use of Amazon Lex is not related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. You may not specify a default value for the childDirected field that does not accurately reflect whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA.

If your use of Amazon Lex relates to a website, program, or other application that is directed in whole or in part, to children under age 13, you must obtain any required verifiable parental consent under COPPA. For information regarding the use of Amazon Lex in connection with websites, programs, or other applications that are directed or targeted, in whole or in part, to children under age 13, see the Amazon Lex FAQ.

$sel:description:PutBot', putBotResponse_description - A description of the bot.

$sel:createVersion:PutBot', putBotResponse_createVersion - True if a new version of the bot was created. If the createVersion field was not specified in the request, the createVersion field is set to false in the response.

$sel:detectSentiment:PutBot', putBotResponse_detectSentiment - true if the bot is configured to send user utterances to Amazon Comprehend for sentiment analysis. If the detectSentiment field was not specified in the request, the detectSentiment field is false in the response.

$sel:checksum:PutBot', putBotResponse_checksum - Checksum of the bot that you created.

$sel:httpStatus:PutBotResponse', putBotResponse_httpStatus - The response's http status code.

Response Lenses

putBotResponse_abortStatement :: Lens' PutBotResponse (Maybe Statement)

The message that Amazon Lex uses to cancel a conversation. For more information, see PutBot.

putBotResponse_createdDate :: Lens' PutBotResponse (Maybe UTCTime)

The date that the bot was created.

putBotResponse_status :: Lens' PutBotResponse (Maybe LexStatus)

When you send a request to create a bot with processBehavior set to BUILD, Amazon Lex sets the status response element to BUILDING.

In the READY_BASIC_TESTING state you can test the bot with user inputs that exactly match the utterances configured for the bot's intents and values in the slot types.

If Amazon Lex can't build the bot, Amazon Lex sets status to FAILED. Amazon Lex returns the reason for the failure in the failureReason response element.

When you set processBehavior to SAVE, Amazon Lex sets the status code to NOT BUILT.

When the bot is in the READY state you can test and publish the bot.

putBotResponse_lastUpdatedDate :: Lens' PutBotResponse (Maybe UTCTime)

The date that the bot was updated. When you create a resource, the creation date and last updated date are the same.

putBotResponse_voiceId :: Lens' PutBotResponse (Maybe Text)

The Amazon Polly voice ID that Amazon Lex uses for voice interaction with the user. For more information, see PutBot.

putBotResponse_clarificationPrompt :: Lens' PutBotResponse (Maybe Prompt)

The prompts that Amazon Lex uses when it doesn't understand the user's intent. For more information, see PutBot.

putBotResponse_nluIntentConfidenceThreshold :: Lens' PutBotResponse (Maybe Double)

The score that determines where Amazon Lex inserts the AMAZON.FallbackIntent, AMAZON.KendraSearchIntent, or both when returning alternative intents in a PostContent or PostText response. AMAZON.FallbackIntent is inserted if the confidence score for all intents is below this value. AMAZON.KendraSearchIntent is only inserted if it is configured for the bot.

putBotResponse_locale :: Lens' PutBotResponse (Maybe Locale)

The target locale for the bot.

putBotResponse_enableModelImprovements :: Lens' PutBotResponse (Maybe Bool)

Indicates whether the bot uses accuracy improvements. true indicates that the bot is using the improvements, otherwise, false.

putBotResponse_version :: Lens' PutBotResponse (Maybe Text)

The version of the bot. For a new bot, the version is always $LATEST.

putBotResponse_idleSessionTTLInSeconds :: Lens' PutBotResponse (Maybe Natural)

The maximum length of time that Amazon Lex retains the data gathered in a conversation. For more information, see PutBot.

putBotResponse_name :: Lens' PutBotResponse (Maybe Text)

The name of the bot.

putBotResponse_intents :: Lens' PutBotResponse (Maybe [Intent])

An array of Intent objects. For more information, see PutBot.

putBotResponse_failureReason :: Lens' PutBotResponse (Maybe Text)

If status is FAILED, Amazon Lex provides the reason that it failed to build the bot.

putBotResponse_tags :: Lens' PutBotResponse (Maybe [Tag])

A list of tags associated with the bot.

putBotResponse_childDirected :: Lens' PutBotResponse (Maybe Bool)

For each Amazon Lex bot created with the Amazon Lex Model Building Service, you must specify whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to the Children's Online Privacy Protection Act (COPPA) by specifying true or false in the childDirected field. By specifying true in the childDirected field, you confirm that your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. By specifying false in the childDirected field, you confirm that your use of Amazon Lex is not related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA. You may not specify a default value for the childDirected field that does not accurately reflect whether your use of Amazon Lex is related to a website, program, or other application that is directed or targeted, in whole or in part, to children under age 13 and subject to COPPA.

If your use of Amazon Lex relates to a website, program, or other application that is directed in whole or in part, to children under age 13, you must obtain any required verifiable parental consent under COPPA. For information regarding the use of Amazon Lex in connection with websites, programs, or other applications that are directed or targeted, in whole or in part, to children under age 13, see the Amazon Lex FAQ.

putBotResponse_description :: Lens' PutBotResponse (Maybe Text)

A description of the bot.

putBotResponse_createVersion :: Lens' PutBotResponse (Maybe Bool)

True if a new version of the bot was created. If the createVersion field was not specified in the request, the createVersion field is set to false in the response.

putBotResponse_detectSentiment :: Lens' PutBotResponse (Maybe Bool)

true if the bot is configured to send user utterances to Amazon Comprehend for sentiment analysis. If the detectSentiment field was not specified in the request, the detectSentiment field is false in the response.

putBotResponse_checksum :: Lens' PutBotResponse (Maybe Text)

Checksum of the bot that you created.

putBotResponse_httpStatus :: Lens' PutBotResponse Int

The response's http status code.