Network.AWS.LexModels.PutBotAlias

Description

Creates an alias for the specified version of the bot or replaces an alias for the specified bot. To change the version of the bot that the alias points to, replace the alias. For more information about aliases, see versioning-aliases.

This operation requires permissions for the lex:PutBotAlias action.

Creating a Request

data PutBotAlias

See: newPutBotAlias smart constructor.

Constructors

PutBotAlias'

Fields

checksum :: Maybe Text

Identifies a specific revision of the $LATEST version.

When you create a new bot alias, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception.

When you want to update a bot alias, set the checksum field to the checksum of the most recent revision of the $LATEST version. If you don't specify the checksum field, or if the checksum does not match the $LATEST version, you get a PreconditionFailedException exception.

conversationLogs :: Maybe ConversationLogsRequest

Settings for conversation logs for the alias.

description :: Maybe Text

A description of the alias.

tags :: Maybe [Tag]

A list of tags to add to the bot alias. You can only add tags when you create an alias, you can't use the PutBotAlias operation to update the tags on a bot alias. To update tags, use the TagResource operation.

name :: Text

The name of the alias. The name is not case sensitive.

botVersion :: Text

The version of the bot.

botName :: Text

The name of the bot.

newPutBotAlias

Arguments

:: Text
$sel:name:PutBotAlias'

-> Text
$sel:botVersion:PutBotAlias'

-> Text
$sel:botName:PutBotAlias'

-> PutBotAlias

Create a value of PutBotAlias with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checksum:PutBotAlias', putBotAlias_checksum - Identifies a specific revision of the $LATEST version.

When you create a new bot alias, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception.

When you want to update a bot alias, set the checksum field to the checksum of the most recent revision of the $LATEST version. If you don't specify the checksum field, or if the checksum does not match the $LATEST version, you get a PreconditionFailedException exception.

$sel:conversationLogs:PutBotAlias', putBotAlias_conversationLogs - Settings for conversation logs for the alias.

$sel:description:PutBotAlias', putBotAlias_description - A description of the alias.

$sel:tags:PutBotAlias', putBotAlias_tags - A list of tags to add to the bot alias. You can only add tags when you create an alias, you can't use the PutBotAlias operation to update the tags on a bot alias. To update tags, use the TagResource operation.

$sel:name:PutBotAlias', putBotAlias_name - The name of the alias. The name is not case sensitive.

$sel:botVersion:PutBotAlias', putBotAlias_botVersion - The version of the bot.

$sel:botName:PutBotAlias', putBotAlias_botName - The name of the bot.

Request Lenses

putBotAlias_checksum :: Lens' PutBotAlias (Maybe Text)

Identifies a specific revision of the $LATEST version.

When you create a new bot alias, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception.

When you want to update a bot alias, set the checksum field to the checksum of the most recent revision of the $LATEST version. If you don't specify the checksum field, or if the checksum does not match the $LATEST version, you get a PreconditionFailedException exception.

putBotAlias_conversationLogs :: Lens' PutBotAlias (Maybe ConversationLogsRequest)

Settings for conversation logs for the alias.

putBotAlias_description :: Lens' PutBotAlias (Maybe Text)

A description of the alias.

putBotAlias_tags :: Lens' PutBotAlias (Maybe [Tag])

A list of tags to add to the bot alias. You can only add tags when you create an alias, you can't use the PutBotAlias operation to update the tags on a bot alias. To update tags, use the TagResource operation.

putBotAlias_name :: Lens' PutBotAlias Text

The name of the alias. The name is not case sensitive.

putBotAlias_botVersion :: Lens' PutBotAlias Text

The version of the bot.

putBotAlias_botName :: Lens' PutBotAlias Text

The name of the bot.

Destructuring the Response

data PutBotAliasResponse

See: newPutBotAliasResponse smart constructor.

Constructors

PutBotAliasResponse'

Fields

checksum :: Maybe Text

The checksum for the current version of the alias.

botVersion :: Maybe Text

The version of the bot that the alias points to.

botName :: Maybe Text

The name of the bot that the alias points to.

createdDate :: Maybe POSIX

The date that the bot alias was created.

name :: Maybe Text

The name of the alias.

conversationLogs :: Maybe ConversationLogsResponse

The settings that determine how Amazon Lex uses conversation logs for the alias.

lastUpdatedDate :: Maybe POSIX

The date that the bot alias was updated. When you create a resource, the creation date and the last updated date are the same.

description :: Maybe Text

A description of the alias.

tags :: Maybe [Tag]

A list of tags associated with a bot.

httpStatus :: Int

The response's http status code.

newPutBotAliasResponse

Arguments

:: Int
$sel:httpStatus:PutBotAliasResponse'

-> PutBotAliasResponse

Create a value of PutBotAliasResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checksum:PutBotAlias', putBotAliasResponse_checksum - The checksum for the current version of the alias.

$sel:botVersion:PutBotAlias', putBotAliasResponse_botVersion - The version of the bot that the alias points to.

$sel:botName:PutBotAlias', putBotAliasResponse_botName - The name of the bot that the alias points to.

$sel:createdDate:PutBotAliasResponse', putBotAliasResponse_createdDate - The date that the bot alias was created.

$sel:name:PutBotAlias', putBotAliasResponse_name - The name of the alias.

$sel:conversationLogs:PutBotAlias', putBotAliasResponse_conversationLogs - The settings that determine how Amazon Lex uses conversation logs for the alias.

$sel:lastUpdatedDate:PutBotAliasResponse', putBotAliasResponse_lastUpdatedDate - The date that the bot alias was updated. When you create a resource, the creation date and the last updated date are the same.

$sel:description:PutBotAlias', putBotAliasResponse_description - A description of the alias.

$sel:tags:PutBotAlias', putBotAliasResponse_tags - A list of tags associated with a bot.

$sel:httpStatus:PutBotAliasResponse', putBotAliasResponse_httpStatus - The response's http status code.

Response Lenses

putBotAliasResponse_checksum :: Lens' PutBotAliasResponse (Maybe Text)

The checksum for the current version of the alias.

putBotAliasResponse_botVersion :: Lens' PutBotAliasResponse (Maybe Text)

The version of the bot that the alias points to.

putBotAliasResponse_botName :: Lens' PutBotAliasResponse (Maybe Text)

The name of the bot that the alias points to.

putBotAliasResponse_createdDate :: Lens' PutBotAliasResponse (Maybe UTCTime)

The date that the bot alias was created.

putBotAliasResponse_name :: Lens' PutBotAliasResponse (Maybe Text)

The name of the alias.

putBotAliasResponse_conversationLogs :: Lens' PutBotAliasResponse (Maybe ConversationLogsResponse)

The settings that determine how Amazon Lex uses conversation logs for the alias.

putBotAliasResponse_lastUpdatedDate :: Lens' PutBotAliasResponse (Maybe UTCTime)

The date that the bot alias was updated. When you create a resource, the creation date and the last updated date are the same.

putBotAliasResponse_description :: Lens' PutBotAliasResponse (Maybe Text)

A description of the alias.

putBotAliasResponse_tags :: Lens' PutBotAliasResponse (Maybe [Tag])

A list of tags associated with a bot.

putBotAliasResponse_httpStatus :: Lens' PutBotAliasResponse Int

The response's http status code.