Network.AWS.LexModels.PutIntent

Description

Creates an intent or replaces an existing intent.

To define the interaction between the user and your bot, you use one or more intents. For a pizza ordering bot, for example, you would create an OrderPizza intent.

To create an intent or replace an existing intent, you must provide the following:

Intent name. For example, OrderPizza.

Sample utterances. For example, "Can I order a pizza, please." and "I want to order a pizza."

Information to be gathered. You specify slot types for the information that your bot will request from the user. You can specify standard slot types, such as a date or a time, or custom slot types such as the size and crust of a pizza.

How the intent will be fulfilled. You can provide a Lambda function or configure the intent to return the intent information to the client application. If you use a Lambda function, when all of the intent information is available, Amazon Lex invokes your Lambda function. If you configure your intent to return the intent information to the client application.

You can specify other optional information in the request, such as:

A confirmation prompt to ask the user to confirm an intent. For example, "Shall I order your pizza?"

A conclusion statement to send to the user after the intent has been fulfilled. For example, "I placed your pizza order."

A follow-up prompt that asks the user for additional activity. For example, asking "Do you want to order a drink with your pizza?"

If you specify an existing intent name to update the intent, Amazon Lex replaces the values in the $LATEST version of the intent with the values in the request. Amazon Lex removes fields that you don't provide in the request. If you don't specify the required fields, Amazon Lex throws an exception. When you update the $LATEST version of an intent, the status field of any bot that uses the $LATEST version of the intent is set to NOT_BUILT.

For more information, see how-it-works.

This operation requires permissions for the lex:PutIntent action.

Creating a Request

data PutIntent

See: newPutIntent smart constructor.

Constructors

PutIntent'

Fields

kendraConfiguration :: Maybe KendraConfiguration

Configuration information required to use the AMAZON.KendraSearchIntent intent to connect to an Amazon Kendra index. For more information, see AMAZON.KendraSearchIntent.

parentIntentSignature :: Maybe Text

A unique identifier for the built-in intent to base this intent on. To find the signature for an intent, see Standard Built-in Intents in the Alexa Skills Kit.

dialogCodeHook :: Maybe CodeHook

Specifies a Lambda function to invoke for each user input. You can invoke this Lambda function to personalize user interaction.

For example, suppose your bot determines that the user is John. Your Lambda function might retrieve John's information from a backend database and prepopulate some of the values. For example, if you find that John is gluten intolerant, you might set the corresponding intent slot, GlutenIntolerant, to true. You might find John's phone number and set the corresponding session attribute.

conclusionStatement :: Maybe Statement

The statement that you want Amazon Lex to convey to the user after the intent is successfully fulfilled by the Lambda function.

This element is relevant only if you provide a Lambda function in the fulfillmentActivity. If you return the intent to the client application, you can't specify this element.

The followUpPrompt and conclusionStatement are mutually exclusive. You can specify only one.

rejectionStatement :: Maybe Statement

When the user answers "no" to the question defined in confirmationPrompt, Amazon Lex responds with this statement to acknowledge that the intent was canceled.

You must provide both the rejectionStatement and the confirmationPrompt, or neither.

inputContexts :: Maybe [InputContext]

An array of InputContext objects that lists the contexts that must be active for Amazon Lex to choose the intent in a conversation with the user.

slots :: Maybe [Slot]

An array of intent slots. At runtime, Amazon Lex elicits required slot values from the user using prompts defined in the slots. For more information, see how-it-works.

fulfillmentActivity :: Maybe FulfillmentActivity

Required. Describes how the intent is fulfilled. For example, after a user provides all of the information for a pizza order, fulfillmentActivity defines how the bot places an order with a local pizza store.

You might configure Amazon Lex to return all of the intent information to the client application, or direct it to invoke a Lambda function that can process the intent (for example, place an order with a pizzeria).

sampleUtterances :: Maybe [Text]

An array of utterances (strings) that a user might say to signal the intent. For example, "I want {PizzaSize} pizza", "Order {Quantity} {PizzaSize} pizzas".

In each utterance, a slot name is enclosed in curly braces.

description :: Maybe Text

A description of the intent.

createVersion :: Maybe Bool

When set to true a new numbered version of the intent is created. This is the same as calling the CreateIntentVersion operation. If you do not specify createVersion, the default is false.

confirmationPrompt :: Maybe Prompt

Prompts the user to confirm the intent. This question should have a yes or no answer.

Amazon Lex uses this prompt to ensure that the user acknowledges that the intent is ready for fulfillment. For example, with the OrderPizza intent, you might want to confirm that the order is correct before placing it. For other intents, such as intents that simply respond to user questions, you might not need to ask the user for confirmation before providing the information.

You you must provide both the rejectionStatement and the confirmationPrompt, or neither.

outputContexts :: Maybe [OutputContext]

An array of OutputContext objects that lists the contexts that the intent activates when the intent is fulfilled.

followUpPrompt :: Maybe FollowUpPrompt

Amazon Lex uses this prompt to solicit additional activity after fulfilling an intent. For example, after the OrderPizza intent is fulfilled, you might prompt the user to order a drink.

The action that Amazon Lex takes depends on the user's response, as follows:

If the user says "Yes" it responds with the clarification prompt that is configured for the bot.

if the user says "Yes" and continues with an utterance that triggers an intent it starts a conversation for the intent.

If the user says "No" it responds with the rejection statement configured for the the follow-up prompt.

If it doesn't recognize the utterance it repeats the follow-up prompt again.

The followUpPrompt field and the conclusionStatement field are mutually exclusive. You can specify only one.

checksum :: Maybe Text

Identifies a specific revision of the $LATEST version.

When you create a new intent, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception.

When you want to update a intent, set the checksum field to the checksum of the most recent revision of the $LATEST version. If you don't specify the checksum field, or if the checksum does not match the $LATEST version, you get a PreconditionFailedException exception.

name :: Text

The name of the intent. The name is not case sensitive.

The name can't match a built-in intent name, or a built-in intent name with "AMAZON." removed. For example, because there is a built-in intent called AMAZON.HelpIntent, you can't create a custom intent called HelpIntent.

For a list of built-in intents, see Standard Built-in Intents in the Alexa Skills Kit.

newPutIntent

Arguments

:: Text
$sel:name:PutIntent'

-> PutIntent

Create a value of PutIntent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kendraConfiguration:PutIntent', putIntent_kendraConfiguration - Configuration information required to use the AMAZON.KendraSearchIntent intent to connect to an Amazon Kendra index. For more information, see AMAZON.KendraSearchIntent.

$sel:parentIntentSignature:PutIntent', putIntent_parentIntentSignature - A unique identifier for the built-in intent to base this intent on. To find the signature for an intent, see Standard Built-in Intents in the Alexa Skills Kit.

$sel:dialogCodeHook:PutIntent', putIntent_dialogCodeHook - Specifies a Lambda function to invoke for each user input. You can invoke this Lambda function to personalize user interaction.

For example, suppose your bot determines that the user is John. Your Lambda function might retrieve John's information from a backend database and prepopulate some of the values. For example, if you find that John is gluten intolerant, you might set the corresponding intent slot, GlutenIntolerant, to true. You might find John's phone number and set the corresponding session attribute.

$sel:conclusionStatement:PutIntent', putIntent_conclusionStatement - The statement that you want Amazon Lex to convey to the user after the intent is successfully fulfilled by the Lambda function.

This element is relevant only if you provide a Lambda function in the fulfillmentActivity. If you return the intent to the client application, you can't specify this element.

The followUpPrompt and conclusionStatement are mutually exclusive. You can specify only one.

$sel:rejectionStatement:PutIntent', putIntent_rejectionStatement - When the user answers "no" to the question defined in confirmationPrompt, Amazon Lex responds with this statement to acknowledge that the intent was canceled.

You must provide both the rejectionStatement and the confirmationPrompt, or neither.

$sel:inputContexts:PutIntent', putIntent_inputContexts - An array of InputContext objects that lists the contexts that must be active for Amazon Lex to choose the intent in a conversation with the user.

$sel:slots:PutIntent', putIntent_slots - An array of intent slots. At runtime, Amazon Lex elicits required slot values from the user using prompts defined in the slots. For more information, see how-it-works.

$sel:fulfillmentActivity:PutIntent', putIntent_fulfillmentActivity - Required. Describes how the intent is fulfilled. For example, after a user provides all of the information for a pizza order, fulfillmentActivity defines how the bot places an order with a local pizza store.

You might configure Amazon Lex to return all of the intent information to the client application, or direct it to invoke a Lambda function that can process the intent (for example, place an order with a pizzeria).

$sel:sampleUtterances:PutIntent', putIntent_sampleUtterances - An array of utterances (strings) that a user might say to signal the intent. For example, "I want {PizzaSize} pizza", "Order {Quantity} {PizzaSize} pizzas".

In each utterance, a slot name is enclosed in curly braces.

$sel:description:PutIntent', putIntent_description - A description of the intent.

$sel:createVersion:PutIntent', putIntent_createVersion - When set to true a new numbered version of the intent is created. This is the same as calling the CreateIntentVersion operation. If you do not specify createVersion, the default is false.

$sel:confirmationPrompt:PutIntent', putIntent_confirmationPrompt - Prompts the user to confirm the intent. This question should have a yes or no answer.

Amazon Lex uses this prompt to ensure that the user acknowledges that the intent is ready for fulfillment. For example, with the OrderPizza intent, you might want to confirm that the order is correct before placing it. For other intents, such as intents that simply respond to user questions, you might not need to ask the user for confirmation before providing the information.

You you must provide both the rejectionStatement and the confirmationPrompt, or neither.

$sel:outputContexts:PutIntent', putIntent_outputContexts - An array of OutputContext objects that lists the contexts that the intent activates when the intent is fulfilled.

$sel:followUpPrompt:PutIntent', putIntent_followUpPrompt - Amazon Lex uses this prompt to solicit additional activity after fulfilling an intent. For example, after the OrderPizza intent is fulfilled, you might prompt the user to order a drink.

The action that Amazon Lex takes depends on the user's response, as follows:

If the user says "Yes" it responds with the clarification prompt that is configured for the bot.

if the user says "Yes" and continues with an utterance that triggers an intent it starts a conversation for the intent.

If the user says "No" it responds with the rejection statement configured for the the follow-up prompt.

If it doesn't recognize the utterance it repeats the follow-up prompt again.

The followUpPrompt field and the conclusionStatement field are mutually exclusive. You can specify only one.

$sel:checksum:PutIntent', putIntent_checksum - Identifies a specific revision of the $LATEST version.

When you create a new intent, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception.

When you want to update a intent, set the checksum field to the checksum of the most recent revision of the $LATEST version. If you don't specify the checksum field, or if the checksum does not match the $LATEST version, you get a PreconditionFailedException exception.

$sel:name:PutIntent', putIntent_name - The name of the intent. The name is not case sensitive.

The name can't match a built-in intent name, or a built-in intent name with "AMAZON." removed. For example, because there is a built-in intent called AMAZON.HelpIntent, you can't create a custom intent called HelpIntent.

For a list of built-in intents, see Standard Built-in Intents in the Alexa Skills Kit.

Request Lenses

putIntent_kendraConfiguration :: Lens' PutIntent (Maybe KendraConfiguration)

Configuration information required to use the AMAZON.KendraSearchIntent intent to connect to an Amazon Kendra index. For more information, see AMAZON.KendraSearchIntent.

putIntent_parentIntentSignature :: Lens' PutIntent (Maybe Text)

A unique identifier for the built-in intent to base this intent on. To find the signature for an intent, see Standard Built-in Intents in the Alexa Skills Kit.

putIntent_dialogCodeHook :: Lens' PutIntent (Maybe CodeHook)

Specifies a Lambda function to invoke for each user input. You can invoke this Lambda function to personalize user interaction.

For example, suppose your bot determines that the user is John. Your Lambda function might retrieve John's information from a backend database and prepopulate some of the values. For example, if you find that John is gluten intolerant, you might set the corresponding intent slot, GlutenIntolerant, to true. You might find John's phone number and set the corresponding session attribute.

putIntent_conclusionStatement :: Lens' PutIntent (Maybe Statement)

The statement that you want Amazon Lex to convey to the user after the intent is successfully fulfilled by the Lambda function.

This element is relevant only if you provide a Lambda function in the fulfillmentActivity. If you return the intent to the client application, you can't specify this element.

The followUpPrompt and conclusionStatement are mutually exclusive. You can specify only one.

putIntent_rejectionStatement :: Lens' PutIntent (Maybe Statement)

When the user answers "no" to the question defined in confirmationPrompt, Amazon Lex responds with this statement to acknowledge that the intent was canceled.

You must provide both the rejectionStatement and the confirmationPrompt, or neither.

putIntent_inputContexts :: Lens' PutIntent (Maybe [InputContext])

An array of InputContext objects that lists the contexts that must be active for Amazon Lex to choose the intent in a conversation with the user.

putIntent_slots :: Lens' PutIntent (Maybe [Slot])

An array of intent slots. At runtime, Amazon Lex elicits required slot values from the user using prompts defined in the slots. For more information, see how-it-works.

putIntent_fulfillmentActivity :: Lens' PutIntent (Maybe FulfillmentActivity)

Required. Describes how the intent is fulfilled. For example, after a user provides all of the information for a pizza order, fulfillmentActivity defines how the bot places an order with a local pizza store.

You might configure Amazon Lex to return all of the intent information to the client application, or direct it to invoke a Lambda function that can process the intent (for example, place an order with a pizzeria).

putIntent_sampleUtterances :: Lens' PutIntent (Maybe [Text])

An array of utterances (strings) that a user might say to signal the intent. For example, "I want {PizzaSize} pizza", "Order {Quantity} {PizzaSize} pizzas".

In each utterance, a slot name is enclosed in curly braces.

putIntent_description :: Lens' PutIntent (Maybe Text)

A description of the intent.

putIntent_createVersion :: Lens' PutIntent (Maybe Bool)

When set to true a new numbered version of the intent is created. This is the same as calling the CreateIntentVersion operation. If you do not specify createVersion, the default is false.

putIntent_confirmationPrompt :: Lens' PutIntent (Maybe Prompt)

Prompts the user to confirm the intent. This question should have a yes or no answer.

Amazon Lex uses this prompt to ensure that the user acknowledges that the intent is ready for fulfillment. For example, with the OrderPizza intent, you might want to confirm that the order is correct before placing it. For other intents, such as intents that simply respond to user questions, you might not need to ask the user for confirmation before providing the information.

You you must provide both the rejectionStatement and the confirmationPrompt, or neither.

putIntent_outputContexts :: Lens' PutIntent (Maybe [OutputContext])

An array of OutputContext objects that lists the contexts that the intent activates when the intent is fulfilled.

putIntent_followUpPrompt :: Lens' PutIntent (Maybe FollowUpPrompt)

Amazon Lex uses this prompt to solicit additional activity after fulfilling an intent. For example, after the OrderPizza intent is fulfilled, you might prompt the user to order a drink.

The action that Amazon Lex takes depends on the user's response, as follows:

If the user says "Yes" it responds with the clarification prompt that is configured for the bot.

if the user says "Yes" and continues with an utterance that triggers an intent it starts a conversation for the intent.

If the user says "No" it responds with the rejection statement configured for the the follow-up prompt.

If it doesn't recognize the utterance it repeats the follow-up prompt again.

The followUpPrompt field and the conclusionStatement field are mutually exclusive. You can specify only one.

putIntent_checksum :: Lens' PutIntent (Maybe Text)

Identifies a specific revision of the $LATEST version.

When you create a new intent, leave the checksum field blank. If you specify a checksum you get a BadRequestException exception.

When you want to update a intent, set the checksum field to the checksum of the most recent revision of the $LATEST version. If you don't specify the checksum field, or if the checksum does not match the $LATEST version, you get a PreconditionFailedException exception.

putIntent_name :: Lens' PutIntent Text

The name of the intent. The name is not case sensitive.

The name can't match a built-in intent name, or a built-in intent name with "AMAZON." removed. For example, because there is a built-in intent called AMAZON.HelpIntent, you can't create a custom intent called HelpIntent.

For a list of built-in intents, see Standard Built-in Intents in the Alexa Skills Kit.

Destructuring the Response

data PutIntentResponse

See: newPutIntentResponse smart constructor.

Constructors

PutIntentResponse'

Fields

kendraConfiguration :: Maybe KendraConfiguration

Configuration information, if any, required to connect to an Amazon Kendra index and use the AMAZON.KendraSearchIntent intent.

createdDate :: Maybe POSIX

The date that the intent was created.

parentIntentSignature :: Maybe Text

A unique identifier for the built-in intent that this intent is based on.

dialogCodeHook :: Maybe CodeHook

If defined in the intent, Amazon Lex invokes this Lambda function for each user input.

conclusionStatement :: Maybe Statement

After the Lambda function specified in thefulfillmentActivityintent fulfills the intent, Amazon Lex conveys this statement to the user.

lastUpdatedDate :: Maybe POSIX

The date that the intent was updated. When you create a resource, the creation date and last update dates are the same.

rejectionStatement :: Maybe Statement

If the user answers "no" to the question defined in confirmationPrompt Amazon Lex responds with this statement to acknowledge that the intent was canceled.

inputContexts :: Maybe [InputContext]

An array of InputContext objects that lists the contexts that must be active for Amazon Lex to choose the intent in a conversation with the user.

version :: Maybe Text

The version of the intent. For a new intent, the version is always $LATEST.

slots :: Maybe [Slot]

An array of intent slots that are configured for the intent.

fulfillmentActivity :: Maybe FulfillmentActivity

If defined in the intent, Amazon Lex invokes this Lambda function to fulfill the intent after the user provides all of the information required by the intent.

name :: Maybe Text

The name of the intent.

sampleUtterances :: Maybe [Text]

An array of sample utterances that are configured for the intent.

description :: Maybe Text

A description of the intent.

createVersion :: Maybe Bool

True if a new version of the intent was created. If the createVersion field was not specified in the request, the createVersion field is set to false in the response.

confirmationPrompt :: Maybe Prompt

If defined in the intent, Amazon Lex prompts the user to confirm the intent before fulfilling it.

outputContexts :: Maybe [OutputContext]

An array of OutputContext objects that lists the contexts that the intent activates when the intent is fulfilled.

followUpPrompt :: Maybe FollowUpPrompt

If defined in the intent, Amazon Lex uses this prompt to solicit additional user activity after the intent is fulfilled.

checksum :: Maybe Text

Checksum of the $LATESTversion of the intent created or updated.

httpStatus :: Int

The response's http status code.

newPutIntentResponse

Arguments

:: Int
$sel:httpStatus:PutIntentResponse'

-> PutIntentResponse

Create a value of PutIntentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kendraConfiguration:PutIntent', putIntentResponse_kendraConfiguration - Configuration information, if any, required to connect to an Amazon Kendra index and use the AMAZON.KendraSearchIntent intent.

$sel:createdDate:PutIntentResponse', putIntentResponse_createdDate - The date that the intent was created.

$sel:parentIntentSignature:PutIntent', putIntentResponse_parentIntentSignature - A unique identifier for the built-in intent that this intent is based on.

$sel:dialogCodeHook:PutIntent', putIntentResponse_dialogCodeHook - If defined in the intent, Amazon Lex invokes this Lambda function for each user input.

$sel:conclusionStatement:PutIntent', putIntentResponse_conclusionStatement - After the Lambda function specified in thefulfillmentActivityintent fulfills the intent, Amazon Lex conveys this statement to the user.

$sel:lastUpdatedDate:PutIntentResponse', putIntentResponse_lastUpdatedDate - The date that the intent was updated. When you create a resource, the creation date and last update dates are the same.

$sel:rejectionStatement:PutIntent', putIntentResponse_rejectionStatement - If the user answers "no" to the question defined in confirmationPrompt Amazon Lex responds with this statement to acknowledge that the intent was canceled.

$sel:inputContexts:PutIntent', putIntentResponse_inputContexts - An array of InputContext objects that lists the contexts that must be active for Amazon Lex to choose the intent in a conversation with the user.

$sel:version:PutIntentResponse', putIntentResponse_version - The version of the intent. For a new intent, the version is always $LATEST.

$sel:slots:PutIntent', putIntentResponse_slots - An array of intent slots that are configured for the intent.

$sel:fulfillmentActivity:PutIntent', putIntentResponse_fulfillmentActivity - If defined in the intent, Amazon Lex invokes this Lambda function to fulfill the intent after the user provides all of the information required by the intent.

$sel:name:PutIntent', putIntentResponse_name - The name of the intent.

$sel:sampleUtterances:PutIntent', putIntentResponse_sampleUtterances - An array of sample utterances that are configured for the intent.

$sel:description:PutIntent', putIntentResponse_description - A description of the intent.

$sel:createVersion:PutIntent', putIntentResponse_createVersion - True if a new version of the intent was created. If the createVersion field was not specified in the request, the createVersion field is set to false in the response.

$sel:confirmationPrompt:PutIntent', putIntentResponse_confirmationPrompt - If defined in the intent, Amazon Lex prompts the user to confirm the intent before fulfilling it.

$sel:outputContexts:PutIntent', putIntentResponse_outputContexts - An array of OutputContext objects that lists the contexts that the intent activates when the intent is fulfilled.

$sel:followUpPrompt:PutIntent', putIntentResponse_followUpPrompt - If defined in the intent, Amazon Lex uses this prompt to solicit additional user activity after the intent is fulfilled.

$sel:checksum:PutIntent', putIntentResponse_checksum - Checksum of the $LATESTversion of the intent created or updated.

$sel:httpStatus:PutIntentResponse', putIntentResponse_httpStatus - The response's http status code.

Response Lenses

putIntentResponse_kendraConfiguration :: Lens' PutIntentResponse (Maybe KendraConfiguration)

Configuration information, if any, required to connect to an Amazon Kendra index and use the AMAZON.KendraSearchIntent intent.

putIntentResponse_createdDate :: Lens' PutIntentResponse (Maybe UTCTime)

The date that the intent was created.

putIntentResponse_parentIntentSignature :: Lens' PutIntentResponse (Maybe Text)

A unique identifier for the built-in intent that this intent is based on.

putIntentResponse_dialogCodeHook :: Lens' PutIntentResponse (Maybe CodeHook)

If defined in the intent, Amazon Lex invokes this Lambda function for each user input.

putIntentResponse_conclusionStatement :: Lens' PutIntentResponse (Maybe Statement)

After the Lambda function specified in thefulfillmentActivityintent fulfills the intent, Amazon Lex conveys this statement to the user.

putIntentResponse_lastUpdatedDate :: Lens' PutIntentResponse (Maybe UTCTime)

The date that the intent was updated. When you create a resource, the creation date and last update dates are the same.

putIntentResponse_rejectionStatement :: Lens' PutIntentResponse (Maybe Statement)

If the user answers "no" to the question defined in confirmationPrompt Amazon Lex responds with this statement to acknowledge that the intent was canceled.

putIntentResponse_inputContexts :: Lens' PutIntentResponse (Maybe [InputContext])

An array of InputContext objects that lists the contexts that must be active for Amazon Lex to choose the intent in a conversation with the user.

putIntentResponse_version :: Lens' PutIntentResponse (Maybe Text)

The version of the intent. For a new intent, the version is always $LATEST.

putIntentResponse_slots :: Lens' PutIntentResponse (Maybe [Slot])

An array of intent slots that are configured for the intent.

putIntentResponse_fulfillmentActivity :: Lens' PutIntentResponse (Maybe FulfillmentActivity)

If defined in the intent, Amazon Lex invokes this Lambda function to fulfill the intent after the user provides all of the information required by the intent.

putIntentResponse_name :: Lens' PutIntentResponse (Maybe Text)

The name of the intent.

putIntentResponse_sampleUtterances :: Lens' PutIntentResponse (Maybe [Text])

An array of sample utterances that are configured for the intent.

putIntentResponse_description :: Lens' PutIntentResponse (Maybe Text)

A description of the intent.

putIntentResponse_createVersion :: Lens' PutIntentResponse (Maybe Bool)

True if a new version of the intent was created. If the createVersion field was not specified in the request, the createVersion field is set to false in the response.

putIntentResponse_confirmationPrompt :: Lens' PutIntentResponse (Maybe Prompt)

If defined in the intent, Amazon Lex prompts the user to confirm the intent before fulfilling it.

putIntentResponse_outputContexts :: Lens' PutIntentResponse (Maybe [OutputContext])

An array of OutputContext objects that lists the contexts that the intent activates when the intent is fulfilled.

putIntentResponse_followUpPrompt :: Lens' PutIntentResponse (Maybe FollowUpPrompt)

If defined in the intent, Amazon Lex uses this prompt to solicit additional user activity after the intent is fulfilled.

putIntentResponse_checksum :: Lens' PutIntentResponse (Maybe Text)

Checksum of the $LATESTversion of the intent created or updated.

putIntentResponse_httpStatus :: Lens' PutIntentResponse Int

The response's http status code.