Network.AWS.LexModels.Types.BotChannelAssociation

Description

Documentation

data BotChannelAssociation

Represents an association between an Amazon Lex bot and an external messaging platform.

See: newBotChannelAssociation smart constructor.

Constructors

BotChannelAssociation'

Fields

botAlias :: Maybe Text

An alias pointing to the specific version of the Amazon Lex bot to which this association is being made.

createdDate :: Maybe POSIX

The date that the association between the Amazon Lex bot and the channel was created.

status :: Maybe ChannelStatus

The status of the bot channel.

CREATED - The channel has been created and is ready for use.

IN_PROGRESS - Channel creation is in progress.

FAILED - There was an error creating the channel. For information about the reason for the failure, see the failureReason field.

botConfiguration :: Maybe (Sensitive (HashMap Text Text))

Provides information necessary to communicate with the messaging platform.

botName :: Maybe Text

The name of the Amazon Lex bot to which this association is being made.

Currently, Amazon Lex supports associations with Facebook and Slack, and Twilio.

name :: Maybe Text

The name of the association between the bot and the channel.

failureReason :: Maybe Text

If status is FAILED, Amazon Lex provides the reason that it failed to create the association.

description :: Maybe Text

A text description of the association you are creating.

type' :: Maybe ChannelType

Specifies the type of association by indicating the type of channel being established between the Amazon Lex bot and the external messaging platform.

newBotChannelAssociation :: BotChannelAssociation

Create a value of BotChannelAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:botAlias:BotChannelAssociation', botChannelAssociation_botAlias - An alias pointing to the specific version of the Amazon Lex bot to which this association is being made.

$sel:createdDate:BotChannelAssociation', botChannelAssociation_createdDate - The date that the association between the Amazon Lex bot and the channel was created.

$sel:status:BotChannelAssociation', botChannelAssociation_status - The status of the bot channel.

CREATED - The channel has been created and is ready for use.

IN_PROGRESS - Channel creation is in progress.

FAILED - There was an error creating the channel. For information about the reason for the failure, see the failureReason field.

$sel:botConfiguration:BotChannelAssociation', botChannelAssociation_botConfiguration - Provides information necessary to communicate with the messaging platform.

$sel:botName:BotChannelAssociation', botChannelAssociation_botName - The name of the Amazon Lex bot to which this association is being made.

Currently, Amazon Lex supports associations with Facebook and Slack, and Twilio.

$sel:name:BotChannelAssociation', botChannelAssociation_name - The name of the association between the bot and the channel.

$sel:failureReason:BotChannelAssociation', botChannelAssociation_failureReason - If status is FAILED, Amazon Lex provides the reason that it failed to create the association.

$sel:description:BotChannelAssociation', botChannelAssociation_description - A text description of the association you are creating.

$sel:type':BotChannelAssociation', botChannelAssociation_type - Specifies the type of association by indicating the type of channel being established between the Amazon Lex bot and the external messaging platform.

botChannelAssociation_botAlias :: Lens' BotChannelAssociation (Maybe Text)

An alias pointing to the specific version of the Amazon Lex bot to which this association is being made.

botChannelAssociation_createdDate :: Lens' BotChannelAssociation (Maybe UTCTime)

The date that the association between the Amazon Lex bot and the channel was created.

botChannelAssociation_status :: Lens' BotChannelAssociation (Maybe ChannelStatus)

The status of the bot channel.

CREATED - The channel has been created and is ready for use.

IN_PROGRESS - Channel creation is in progress.

FAILED - There was an error creating the channel. For information about the reason for the failure, see the failureReason field.

botChannelAssociation_botConfiguration :: Lens' BotChannelAssociation (Maybe (HashMap Text Text))

Provides information necessary to communicate with the messaging platform.

botChannelAssociation_botName :: Lens' BotChannelAssociation (Maybe Text)

The name of the Amazon Lex bot to which this association is being made.

Currently, Amazon Lex supports associations with Facebook and Slack, and Twilio.

botChannelAssociation_name :: Lens' BotChannelAssociation (Maybe Text)

The name of the association between the bot and the channel.

botChannelAssociation_failureReason :: Lens' BotChannelAssociation (Maybe Text)

If status is FAILED, Amazon Lex provides the reason that it failed to create the association.

botChannelAssociation_description :: Lens' BotChannelAssociation (Maybe Text)

A text description of the association you are creating.

botChannelAssociation_type :: Lens' BotChannelAssociation (Maybe ChannelType)

Specifies the type of association by indicating the type of channel being established between the Amazon Lex bot and the external messaging platform.