Network.AWS.LexModels.Types.CodeHook

Description

Documentation

data CodeHook

Specifies a Lambda function that verifies requests to a bot or fulfills the user's request to a bot..

See: newCodeHook smart constructor.

Constructors

CodeHook'

Fields

uri :: Text

The Amazon Resource Name (ARN) of the Lambda function.

messageVersion :: Text

The version of the request-response that you want Amazon Lex to use to invoke your Lambda function. For more information, see using-lambda.

newCodeHook

Arguments

:: Text
$sel:uri:CodeHook'

-> Text
$sel:messageVersion:CodeHook'

-> CodeHook

Create a value of CodeHook with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:uri:CodeHook', codeHook_uri - The Amazon Resource Name (ARN) of the Lambda function.

$sel:messageVersion:CodeHook', codeHook_messageVersion - The version of the request-response that you want Amazon Lex to use to invoke your Lambda function. For more information, see using-lambda.

codeHook_uri :: Lens' CodeHook Text

The Amazon Resource Name (ARN) of the Lambda function.

codeHook_messageVersion :: Lens' CodeHook Text

The version of the request-response that you want Amazon Lex to use to invoke your Lambda function. For more information, see using-lambda.