Network.AWS.LexModels.Types.Intent

Description

Documentation

data Intent

Identifies the specific version of an intent.

See: newIntent smart constructor.

Constructors

Intent'

Fields

intentName :: Text

The name of the intent.

intentVersion :: Text

The version of the intent.

newIntent

Arguments

:: Text
$sel:intentName:Intent'

-> Text
$sel:intentVersion:Intent'

-> Intent

Create a value of Intent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:intentName:Intent', intent_intentName - The name of the intent.

$sel:intentVersion:Intent', intent_intentVersion - The version of the intent.

intent_intentName :: Lens' Intent Text

The name of the intent.

intent_intentVersion :: Lens' Intent Text

The version of the intent.