Network.AWS.LexModels.Types.MigrationAlert

Description

Documentation

data MigrationAlert

Provides information about alerts and warnings that Amazon Lex sends during a migration. The alerts include information about how to resolve the issue.

See: newMigrationAlert smart constructor.

Constructors

MigrationAlert'

Fields

message :: Maybe Text

A message that describes why the alert was issued.

details :: Maybe [Text]

Additional details about the alert.

referenceURLs :: Maybe [Text]

A link to the Amazon Lex documentation that describes how to resolve the alert.

type' :: Maybe MigrationAlertType

The type of alert. There are two kinds of alerts:

ERROR - There was an issue with the migration that can't be resolved. The migration stops.

WARN - There was an issue with the migration that requires manual changes to the new Amazon Lex V2 bot. The migration continues.

newMigrationAlert :: MigrationAlert

Create a value of MigrationAlert with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:message:MigrationAlert', migrationAlert_message - A message that describes why the alert was issued.

$sel:details:MigrationAlert', migrationAlert_details - Additional details about the alert.

$sel:referenceURLs:MigrationAlert', migrationAlert_referenceURLs - A link to the Amazon Lex documentation that describes how to resolve the alert.

$sel:type':MigrationAlert', migrationAlert_type - The type of alert. There are two kinds of alerts:

ERROR - There was an issue with the migration that can't be resolved. The migration stops.

WARN - There was an issue with the migration that requires manual changes to the new Amazon Lex V2 bot. The migration continues.

migrationAlert_message :: Lens' MigrationAlert (Maybe Text)

A message that describes why the alert was issued.

migrationAlert_details :: Lens' MigrationAlert (Maybe [Text])

Additional details about the alert.

migrationAlert_referenceURLs :: Lens' MigrationAlert (Maybe [Text])

A link to the Amazon Lex documentation that describes how to resolve the alert.

migrationAlert_type :: Lens' MigrationAlert (Maybe MigrationAlertType)

The type of alert. There are two kinds of alerts:

ERROR - There was an issue with the migration that can't be resolved. The migration stops.

WARN - There was an issue with the migration that requires manual changes to the new Amazon Lex V2 bot. The migration continues.