Network.AWS.LexModels.Types.MigrationSummary

Description

Documentation

data MigrationSummary

Provides information about migrating a bot from Amazon Lex V1 to Amazon Lex V2.

See: newMigrationSummary smart constructor.

Constructors

MigrationSummary'

Fields

v1BotVersion :: Maybe Text

The version of the Amazon Lex V1 bot that is the source of the migration.

migrationId :: Maybe Text

The unique identifier that Amazon Lex assigned to the migration.

v1BotLocale :: Maybe Locale

The locale of the Amazon Lex V1 bot that is the source of the migration.

migrationTimestamp :: Maybe POSIX

The date and time that the migration started.

migrationStrategy :: Maybe MigrationStrategy

The strategy used to conduct the migration.

v2BotRole :: Maybe Text

The IAM role that Amazon Lex uses to run the Amazon Lex V2 bot.

v1BotName :: Maybe Text

The name of the Amazon Lex V1 bot that is the source of the migration.

v2BotId :: Maybe Text

The unique identifier of the Amazon Lex V2 that is the destination of the migration.

migrationStatus :: Maybe MigrationStatus

The status of the operation. When the status is COMPLETE the bot is available in Amazon Lex V2. There may be alerts and warnings that need to be resolved to complete the migration.

newMigrationSummary :: MigrationSummary

Create a value of MigrationSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:v1BotVersion:MigrationSummary', migrationSummary_v1BotVersion - The version of the Amazon Lex V1 bot that is the source of the migration.

$sel:migrationId:MigrationSummary', migrationSummary_migrationId - The unique identifier that Amazon Lex assigned to the migration.

$sel:v1BotLocale:MigrationSummary', migrationSummary_v1BotLocale - The locale of the Amazon Lex V1 bot that is the source of the migration.

$sel:migrationTimestamp:MigrationSummary', migrationSummary_migrationTimestamp - The date and time that the migration started.

$sel:migrationStrategy:MigrationSummary', migrationSummary_migrationStrategy - The strategy used to conduct the migration.

$sel:v2BotRole:MigrationSummary', migrationSummary_v2BotRole - The IAM role that Amazon Lex uses to run the Amazon Lex V2 bot.

$sel:v1BotName:MigrationSummary', migrationSummary_v1BotName - The name of the Amazon Lex V1 bot that is the source of the migration.

$sel:v2BotId:MigrationSummary', migrationSummary_v2BotId - The unique identifier of the Amazon Lex V2 that is the destination of the migration.

$sel:migrationStatus:MigrationSummary', migrationSummary_migrationStatus - The status of the operation. When the status is COMPLETE the bot is available in Amazon Lex V2. There may be alerts and warnings that need to be resolved to complete the migration.

migrationSummary_v1BotVersion :: Lens' MigrationSummary (Maybe Text)

The version of the Amazon Lex V1 bot that is the source of the migration.

migrationSummary_migrationId :: Lens' MigrationSummary (Maybe Text)

The unique identifier that Amazon Lex assigned to the migration.

migrationSummary_v1BotLocale :: Lens' MigrationSummary (Maybe Locale)

The locale of the Amazon Lex V1 bot that is the source of the migration.

migrationSummary_migrationTimestamp :: Lens' MigrationSummary (Maybe UTCTime)

The date and time that the migration started.

migrationSummary_migrationStrategy :: Lens' MigrationSummary (Maybe MigrationStrategy)

The strategy used to conduct the migration.

migrationSummary_v2BotRole :: Lens' MigrationSummary (Maybe Text)

The IAM role that Amazon Lex uses to run the Amazon Lex V2 bot.

migrationSummary_v1BotName :: Lens' MigrationSummary (Maybe Text)

The name of the Amazon Lex V1 bot that is the source of the migration.

migrationSummary_v2BotId :: Lens' MigrationSummary (Maybe Text)

The unique identifier of the Amazon Lex V2 that is the destination of the migration.

migrationSummary_migrationStatus :: Lens' MigrationSummary (Maybe MigrationStatus)

The status of the operation. When the status is COMPLETE the bot is available in Amazon Lex V2. There may be alerts and warnings that need to be resolved to complete the migration.