Network.AWS.LexModels.Types.Slot

Description

Documentation

data Slot

Identifies the version of a specific slot.

See: newSlot smart constructor.

Constructors

Slot'

Fields

responseCard :: Maybe Text

A set of possible responses for the slot type used by text-based clients. A user chooses an option from the response card, instead of using text to reply.

slotType :: Maybe Text

The type of the slot, either a custom slot type that you defined or one of the built-in slot types.

valueElicitationPrompt :: Maybe Prompt

The prompt that Amazon Lex uses to elicit the slot value from the user.

slotTypeVersion :: Maybe Text

The version of the slot type.

priority :: Maybe Natural

Directs Amazon Lex the order in which to elicit this slot value from the user. For example, if the intent has two slots with priorities 1 and 2, AWS Amazon Lex first elicits a value for the slot with priority 1.

If multiple slots share the same priority, the order in which Amazon Lex elicits values is arbitrary.

sampleUtterances :: Maybe [Text]

If you know a specific pattern with which users might respond to an Amazon Lex request for a slot value, you can provide those utterances to improve accuracy. This is optional. In most cases, Amazon Lex is capable of understanding user utterances.

description :: Maybe Text

A description of the slot.

defaultValueSpec :: Maybe SlotDefaultValueSpec

A list of default values for the slot. Default values are used when Amazon Lex hasn't determined a value for a slot. You can specify default values from context variables, session attributes, and defined values.

obfuscationSetting :: Maybe ObfuscationSetting

Determines whether a slot is obfuscated in conversation logs and stored utterances. When you obfuscate a slot, the value is replaced by the slot name in curly braces ({}). For example, if the slot name is "full_name", obfuscated values are replaced with "{full_name}". For more information, see Slot Obfuscation .

name :: Text

The name of the slot.

slotConstraint :: SlotConstraint

Specifies whether the slot is required or optional.

newSlot

Arguments

:: Text
$sel:name:Slot'

-> SlotConstraint
$sel:slotConstraint:Slot'

-> Slot

Create a value of Slot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:responseCard:Slot', slot_responseCard - A set of possible responses for the slot type used by text-based clients. A user chooses an option from the response card, instead of using text to reply.

$sel:slotType:Slot', slot_slotType - The type of the slot, either a custom slot type that you defined or one of the built-in slot types.

$sel:valueElicitationPrompt:Slot', slot_valueElicitationPrompt - The prompt that Amazon Lex uses to elicit the slot value from the user.

$sel:slotTypeVersion:Slot', slot_slotTypeVersion - The version of the slot type.

$sel:priority:Slot', slot_priority - Directs Amazon Lex the order in which to elicit this slot value from the user. For example, if the intent has two slots with priorities 1 and 2, AWS Amazon Lex first elicits a value for the slot with priority 1.

If multiple slots share the same priority, the order in which Amazon Lex elicits values is arbitrary.

$sel:sampleUtterances:Slot', slot_sampleUtterances - If you know a specific pattern with which users might respond to an Amazon Lex request for a slot value, you can provide those utterances to improve accuracy. This is optional. In most cases, Amazon Lex is capable of understanding user utterances.

$sel:description:Slot', slot_description - A description of the slot.

$sel:defaultValueSpec:Slot', slot_defaultValueSpec - A list of default values for the slot. Default values are used when Amazon Lex hasn't determined a value for a slot. You can specify default values from context variables, session attributes, and defined values.

$sel:obfuscationSetting:Slot', slot_obfuscationSetting - Determines whether a slot is obfuscated in conversation logs and stored utterances. When you obfuscate a slot, the value is replaced by the slot name in curly braces ({}). For example, if the slot name is "full_name", obfuscated values are replaced with "{full_name}". For more information, see Slot Obfuscation .

$sel:name:Slot', slot_name - The name of the slot.

$sel:slotConstraint:Slot', slot_slotConstraint - Specifies whether the slot is required or optional.

slot_responseCard :: Lens' Slot (Maybe Text)

A set of possible responses for the slot type used by text-based clients. A user chooses an option from the response card, instead of using text to reply.

slot_slotType :: Lens' Slot (Maybe Text)

The type of the slot, either a custom slot type that you defined or one of the built-in slot types.

slot_valueElicitationPrompt :: Lens' Slot (Maybe Prompt)

The prompt that Amazon Lex uses to elicit the slot value from the user.

slot_slotTypeVersion :: Lens' Slot (Maybe Text)

The version of the slot type.

slot_priority :: Lens' Slot (Maybe Natural)

Directs Amazon Lex the order in which to elicit this slot value from the user. For example, if the intent has two slots with priorities 1 and 2, AWS Amazon Lex first elicits a value for the slot with priority 1.

If multiple slots share the same priority, the order in which Amazon Lex elicits values is arbitrary.

slot_sampleUtterances :: Lens' Slot (Maybe [Text])

If you know a specific pattern with which users might respond to an Amazon Lex request for a slot value, you can provide those utterances to improve accuracy. This is optional. In most cases, Amazon Lex is capable of understanding user utterances.

slot_description :: Lens' Slot (Maybe Text)

A description of the slot.

slot_defaultValueSpec :: Lens' Slot (Maybe SlotDefaultValueSpec)

A list of default values for the slot. Default values are used when Amazon Lex hasn't determined a value for a slot. You can specify default values from context variables, session attributes, and defined values.

slot_obfuscationSetting :: Lens' Slot (Maybe ObfuscationSetting)

Determines whether a slot is obfuscated in conversation logs and stored utterances. When you obfuscate a slot, the value is replaced by the slot name in curly braces ({}). For example, if the slot name is "full_name", obfuscated values are replaced with "{full_name}". For more information, see Slot Obfuscation .

slot_name :: Lens' Slot Text

The name of the slot.

slot_slotConstraint :: Lens' Slot SlotConstraint

Specifies whether the slot is required or optional.