Network.AWS.LexRuntime

Description

Derived from API version 2016-11-28 of the AWS service descriptions, licensed under Apache 2.0.

Amazon Lex provides both build and runtime endpoints. Each endpoint provides a set of operations (API). Your conversational bot uses the runtime API to understand user utterances (user input text or voice). For example, suppose a user says "I want pizza", your bot sends this input to Amazon Lex using the runtime API. Amazon Lex recognizes that the user request is for the OrderPizza intent (one of the intents defined in the bot). Then Amazon Lex engages in user conversation on behalf of the bot to elicit required information (slot values, such as pizza size and crust type), and then performs fulfillment activity (that you configured when you created the bot). You use the build-time API to create and manage your Amazon Lex bot. For a list of build-time operations, see the build-time API, .

Service Configuration

defaultService :: Service

API version 2016-11-28 of the Amazon Lex Runtime Service SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by LexRuntime.

NotAcceptableException

_NotAcceptableException :: AsError a => Getting (First ServiceError) a ServiceError

The accept header in the request does not have a valid value.

DependencyFailedException

_DependencyFailedException :: AsError a => Getting (First ServiceError) a ServiceError

One of the dependencies, such as AWS Lambda or Amazon Polly, threw an exception. For example,

If Amazon Lex does not have sufficient permissions to call a Lambda function.

If a Lambda function takes longer than 30 seconds to execute.

If a fulfillment Lambda function returns a Delegate dialog action without removing any slot values.

UnsupportedMediaTypeException

_UnsupportedMediaTypeException :: AsError a => Getting (First ServiceError) a ServiceError

The Content-Type header (PostContent API) has an invalid value.

ConflictException

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Two clients are using the same AWS account, Amazon Lex bot, and user ID.

NotFoundException

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The resource (such as the Amazon Lex bot or an alias) that is referred to is not found.

RequestTimeoutException

_RequestTimeoutException :: AsError a => Getting (First ServiceError) a ServiceError

The input speech is too long.

LoopDetectedException

_LoopDetectedException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is not used.

InternalFailureException

_InternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

Internal service error. Retry the call.

BadGatewayException

_BadGatewayException :: AsError a => Getting (First ServiceError) a ServiceError

Either the Amazon Lex bot is still building, or one of the dependent services (Amazon Polly, AWS Lambda) failed with an internal service error.

BadRequestException

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

Request validation failed, there is no usable message in the context, or the bot build failed, is still in progress, or contains unbuilt changes.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

Exceeded a limit.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

PutSession

data PutSession

See: newPutSession smart constructor.

Constructors

PutSession' (Maybe Text) (Maybe (Sensitive [ActiveContext])) (Maybe [IntentSummary]) (Maybe DialogAction) (Maybe (Sensitive (HashMap Text Text))) Text Text Text

newPutSession

Arguments

:: Text
$sel:botName:PutSession'

-> Text
$sel:botAlias:PutSession'

-> Text
$sel:userId:PutSession'

-> PutSession

Create a value of PutSession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accept:PutSession', putSession_accept - The message that Amazon Lex returns in the response can be either text or speech based depending on the value of this field.

If the value is text/plain; charset=utf-8, Amazon Lex returns text in the response.

If the value begins with audio/, Amazon Lex returns speech in the response. Amazon Lex uses Amazon Polly to generate the speech in the configuration that you specify. For example, if you specify audio/mpeg as the value, Amazon Lex returns speech in the MPEG format.

If the value is audio/pcm, the speech is returned as audio/pcm in 16-bit, little endian format.

The following are the accepted values:

audio/mpeg

audio/ogg

audio/pcm

audio/* (defaults to mpeg)

text/plain; charset=utf-8

$sel:activeContexts:PutSession', putSession_activeContexts - A list of contexts active for the request. A context can be activated when a previous intent is fulfilled, or by including the context in the request,

If you don't specify a list of contexts, Amazon Lex will use the current list of contexts for the session. If you specify an empty list, all contexts for the session are cleared.

$sel:recentIntentSummaryView:PutSession', putSession_recentIntentSummaryView - A summary of the recent intents for the bot. You can use the intent summary view to set a checkpoint label on an intent and modify attributes of intents. You can also use it to remove or add intent summary objects to the list.

An intent that you modify or add to the list must make sense for the bot. For example, the intent name must be valid for the bot. You must provide valid values for:

intentName

slot names

slotToElict

If you send the recentIntentSummaryView parameter in a PutSession request, the contents of the new summary view replaces the old summary view. For example, if a GetSession request returns three intents in the summary view and you call PutSession with one intent in the summary view, the next call to GetSession will only return one intent.

$sel:dialogAction:PutSession', putSession_dialogAction - Sets the next action that the bot should take to fulfill the conversation.

$sel:sessionAttributes:PutSession', putSession_sessionAttributes - Map of key/value pairs representing the session-specific context information. It contains application information passed between Amazon Lex and a client application.

$sel:botName:PutSession', putSession_botName - The name of the bot that contains the session data.

$sel:botAlias:PutSession', putSession_botAlias - The alias in use for the bot that contains the session data.

$sel:userId:PutSession', putSession_userId - The ID of the client application user. Amazon Lex uses this to identify a user's conversation with your bot.

data PutSessionResponse

See: newPutSessionResponse smart constructor.

Constructors

PutSessionResponse' (Maybe Text) (Maybe Text) (Maybe (Sensitive Text)) (Maybe DialogState) (Maybe (Sensitive Text)) (Maybe MessageFormatType) (Maybe (Sensitive Text)) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) Int ResponseBody

newPutSessionResponse

Arguments

:: Int
$sel:httpStatus:PutSessionResponse'

-> ResponseBody
$sel:audioStream:PutSessionResponse'

-> PutSessionResponse

Create a value of PutSessionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:slots:PutSessionResponse', putSessionResponse_slots - Map of zero or more intent slots Amazon Lex detected from the user input during the conversation.

Amazon Lex creates a resolution list containing likely values for a slot. The value that it returns is determined by the valueSelectionStrategy selected when the slot type was created or updated. If valueSelectionStrategy is set to ORIGINAL_VALUE, the value provided by the user is returned, if the user value is similar to the slot values. If valueSelectionStrategy is set to TOP_RESOLUTION Amazon Lex returns the first value in the resolution list or, if there is no resolution list, null. If you don't specify a valueSelectionStrategy the default is ORIGINAL_VALUE.

$sel:intentName:PutSessionResponse', putSessionResponse_intentName - The name of the current intent.

$sel:encodedMessage:PutSessionResponse', putSessionResponse_encodedMessage - The next message that should be presented to the user.

The encodedMessage field is base-64 encoded. You must decode the field before you can use the value.

$sel:dialogState:PutSessionResponse', putSessionResponse_dialogState - - ConfirmIntent - Amazon Lex is expecting a "yes" or "no" response to confirm the intent before fulfilling an intent.

ElicitIntent - Amazon Lex wants to elicit the user's intent.

ElicitSlot - Amazon Lex is expecting the value of a slot for the current intent.

Failed - Conveys that the conversation with the user has failed. This can happen for various reasons, including the user does not provide an appropriate response to prompts from the service, or if the Lambda function fails to fulfill the intent.

Fulfilled - Conveys that the Lambda function has sucessfully fulfilled the intent.

ReadyForFulfillment - Conveys that the client has to fulfill the intent.

$sel:activeContexts:PutSession', putSessionResponse_activeContexts - A list of active contexts for the session.

$sel:messageFormat:PutSessionResponse', putSessionResponse_messageFormat - The format of the response message. One of the following values:

PlainText - The message contains plain UTF-8 text.

CustomPayload - The message is a custom format for the client.

SSML - The message contains text formatted for voice output.

Composite - The message contains an escaped JSON object containing one or more messages from the groups that messages were assigned to when the intent was created.

$sel:message:PutSessionResponse', putSessionResponse_message - The next message that should be presented to the user.

You can only use this field in the de-DE, en-AU, en-GB, en-US, es-419, es-ES, es-US, fr-CA, fr-FR, and it-IT locales. In all other locales, the message field is null. You should use the encodedMessage field instead.

$sel:sessionId:PutSessionResponse', putSessionResponse_sessionId - A unique identifier for the session.

$sel:slotToElicit:PutSessionResponse', putSessionResponse_slotToElicit - If the dialogState is ElicitSlot, returns the name of the slot for which Amazon Lex is eliciting a value.

$sel:contentType:PutSessionResponse', putSessionResponse_contentType - Content type as specified in the Accept HTTP header in the request.

$sel:sessionAttributes:PutSession', putSessionResponse_sessionAttributes - Map of key/value pairs representing session-specific context information.

$sel:httpStatus:PutSessionResponse', putSessionResponse_httpStatus - The response's http status code.

$sel:audioStream:PutSessionResponse', putSessionResponse_audioStream - The audio version of the message to convey to the user.

DeleteSession

data DeleteSession

See: newDeleteSession smart constructor.

Constructors

DeleteSession' Text Text Text

newDeleteSession

Arguments

:: Text
$sel:botName:DeleteSession'

-> Text
$sel:botAlias:DeleteSession'

-> Text
$sel:userId:DeleteSession'

-> DeleteSession

Create a value of DeleteSession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:botName:DeleteSession', deleteSession_botName - The name of the bot that contains the session data.

$sel:botAlias:DeleteSession', deleteSession_botAlias - The alias in use for the bot that contains the session data.

$sel:userId:DeleteSession', deleteSession_userId - The identifier of the user associated with the session data.

data DeleteSessionResponse

See: newDeleteSessionResponse smart constructor.

Constructors

DeleteSessionResponse' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) Int

newDeleteSessionResponse

Arguments

:: Int
$sel:httpStatus:DeleteSessionResponse'

-> DeleteSessionResponse

Create a value of DeleteSessionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:botAlias:DeleteSession', deleteSessionResponse_botAlias - The alias in use for the bot associated with the session data.

$sel:botName:DeleteSession', deleteSessionResponse_botName - The name of the bot associated with the session data.

$sel:userId:DeleteSession', deleteSessionResponse_userId - The ID of the client application user.

$sel:sessionId:DeleteSessionResponse', deleteSessionResponse_sessionId - The unique identifier for the session.

$sel:httpStatus:DeleteSessionResponse', deleteSessionResponse_httpStatus - The response's http status code.

PostText

data PostText

See: newPostText smart constructor.

Constructors

PostText' (Maybe (Sensitive [ActiveContext])) (Maybe (Sensitive (HashMap Text Text))) (Maybe (Sensitive (HashMap Text Text))) Text Text Text (Sensitive Text)

newPostText

Arguments

:: Text
$sel:botName:PostText'

-> Text
$sel:botAlias:PostText'

-> Text
$sel:userId:PostText'

-> Text
$sel:inputText:PostText'

-> PostText

Create a value of PostText with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:activeContexts:PostText', postText_activeContexts - A list of contexts active for the request. A context can be activated when a previous intent is fulfilled, or by including the context in the request,

If you don't specify a list of contexts, Amazon Lex will use the current list of contexts for the session. If you specify an empty list, all contexts for the session are cleared.

$sel:requestAttributes:PostText', postText_requestAttributes - Request-specific information passed between Amazon Lex and a client application.

The namespace x-amz-lex: is reserved for special attributes. Don't create any request attributes with the prefix x-amz-lex:.

For more information, see Setting Request Attributes.

$sel:sessionAttributes:PostText', postText_sessionAttributes - Application-specific information passed between Amazon Lex and a client application.

For more information, see Setting Session Attributes.

$sel:botName:PostText', postText_botName - The name of the Amazon Lex bot.

$sel:botAlias:PostText', postText_botAlias - The alias of the Amazon Lex bot.

$sel:userId:PostText', postText_userId - The ID of the client application user. Amazon Lex uses this to identify a user's conversation with your bot. At runtime, each request must contain the userID field.

To decide the user ID to use for your application, consider the following factors.

The userID field must not contain any personally identifiable information of the user, for example, name, personal identification numbers, or other end user personal information.

If you want a user to start a conversation on one device and continue on another device, use a user-specific identifier.

If you want the same user to be able to have two independent conversations on two different devices, choose a device-specific identifier.

A user can't have two independent conversations with two different versions of the same bot. For example, a user can't have a conversation with the PROD and BETA versions of the same bot. If you anticipate that a user will need to have conversation with two different versions, for example, while testing, include the bot alias in the user ID to separate the two conversations.

$sel:inputText:PostText', postText_inputText - The text that the user entered (Amazon Lex interprets this text).

data PostTextResponse

See: newPostTextResponse smart constructor.

Constructors

PostTextResponse' (Maybe SentimentResponse) (Maybe IntentConfidence) (Maybe (Sensitive (HashMap Text Text))) (Maybe ResponseCard) (Maybe Text) (Maybe Text) (Maybe DialogState) (Maybe (Sensitive [ActiveContext])) (Maybe [PredictedIntent]) (Maybe MessageFormatType) (Maybe (Sensitive Text)) (Maybe Text) (Maybe Text) (Maybe (Sensitive (HashMap Text Text))) Int

newPostTextResponse

Arguments

:: Int
$sel:httpStatus:PostTextResponse'

-> PostTextResponse

Create a value of PostTextResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sentimentResponse:PostTextResponse', postTextResponse_sentimentResponse - The sentiment expressed in and utterance.

When the bot is configured to send utterances to Amazon Comprehend for sentiment analysis, this field contains the result of the analysis.

$sel:nluIntentConfidence:PostTextResponse', postTextResponse_nluIntentConfidence - Provides a score that indicates how confident Amazon Lex is that the returned intent is the one that matches the user's intent. The score is between 0.0 and 1.0. For more information, see Confidence Scores.

The score is a relative score, not an absolute score. The score may change based on improvements to Amazon Lex.

$sel:slots:PostTextResponse', postTextResponse_slots - The intent slots that Amazon Lex detected from the user input in the conversation.

Amazon Lex creates a resolution list containing likely values for a slot. The value that it returns is determined by the valueSelectionStrategy selected when the slot type was created or updated. If valueSelectionStrategy is set to ORIGINAL_VALUE, the value provided by the user is returned, if the user value is similar to the slot values. If valueSelectionStrategy is set to TOP_RESOLUTION Amazon Lex returns the first value in the resolution list or, if there is no resolution list, null. If you don't specify a valueSelectionStrategy, the default is ORIGINAL_VALUE.

$sel:responseCard:PostTextResponse', postTextResponse_responseCard - Represents the options that the user has to respond to the current prompt. Response Card can come from the bot configuration (in the Amazon Lex console, choose the settings button next to a slot) or from a code hook (Lambda function).

$sel:intentName:PostTextResponse', postTextResponse_intentName - The current user intent that Amazon Lex is aware of.

$sel:botVersion:PostTextResponse', postTextResponse_botVersion - The version of the bot that responded to the conversation. You can use this information to help determine if one version of a bot is performing better than another version.

$sel:dialogState:PostTextResponse', postTextResponse_dialogState - Identifies the current state of the user interaction. Amazon Lex returns one of the following values as dialogState. The client can optionally use this information to customize the user interface.

ElicitIntent - Amazon Lex wants to elicit user intent.

For example, a user might utter an intent ("I want to order a pizza"). If Amazon Lex cannot infer the user intent from this utterance, it will return this dialogState.

ConfirmIntent - Amazon Lex is expecting a "yes" or "no" response.

For example, Amazon Lex wants user confirmation before fulfilling an intent.

Instead of a simple "yes" or "no," a user might respond with additional information. For example, "yes, but make it thick crust pizza" or "no, I want to order a drink". Amazon Lex can process such additional information (in these examples, update the crust type slot value, or change intent from OrderPizza to OrderDrink).

ElicitSlot - Amazon Lex is expecting a slot value for the current intent.

For example, suppose that in the response Amazon Lex sends this message: "What size pizza would you like?". A user might reply with the slot value (e.g., "medium"). The user might also provide additional information in the response (e.g., "medium thick crust pizza"). Amazon Lex can process such additional information appropriately.

Fulfilled - Conveys that the Lambda function configured for the intent has successfully fulfilled the intent.

ReadyForFulfillment - Conveys that the client has to fulfill the intent.

Failed - Conveys that the conversation with the user failed.

This can happen for various reasons including that the user did not provide an appropriate response to prompts from the service (you can configure how many times Amazon Lex can prompt a user for specific information), or the Lambda function failed to fulfill the intent.

$sel:activeContexts:PostText', postTextResponse_activeContexts - A list of active contexts for the session. A context can be set when an intent is fulfilled or by calling the PostContent, PostText, or PutSession operation.

You can use a context to control the intents that can follow up an intent, or to modify the operation of your application.

$sel:alternativeIntents:PostTextResponse', postTextResponse_alternativeIntents - One to four alternative intents that may be applicable to the user's intent.

Each alternative includes a score that indicates how confident Amazon Lex is that the intent matches the user's intent. The intents are sorted by the confidence score.

$sel:messageFormat:PostTextResponse', postTextResponse_messageFormat - The format of the response message. One of the following values:

PlainText - The message contains plain UTF-8 text.

CustomPayload - The message is a custom format defined by the Lambda function.

SSML - The message contains text formatted for voice output.

Composite - The message contains an escaped JSON object containing one or more messages from the groups that messages were assigned to when the intent was created.

$sel:message:PostTextResponse', postTextResponse_message - The message to convey to the user. The message can come from the bot's configuration or from a Lambda function.

If the intent is not configured with a Lambda function, or if the Lambda function returned Delegate as the dialogAction.type its response, Amazon Lex decides on the next course of action and selects an appropriate message from the bot's configuration based on the current interaction context. For example, if Amazon Lex isn't able to understand user input, it uses a clarification prompt message.

When you create an intent you can assign messages to groups. When messages are assigned to groups Amazon Lex returns one message from each group in the response. The message field is an escaped JSON string containing the messages. For more information about the structure of the JSON string returned, see msg-prompts-formats.

If the Lambda function returns a message, Amazon Lex passes it to the client in its response.

$sel:sessionId:PostTextResponse', postTextResponse_sessionId - A unique identifier for the session.

$sel:slotToElicit:PostTextResponse', postTextResponse_slotToElicit - If the dialogState value is ElicitSlot, returns the name of the slot for which Amazon Lex is eliciting a value.

$sel:sessionAttributes:PostText', postTextResponse_sessionAttributes - A map of key-value pairs representing the session-specific context information.

$sel:httpStatus:PostTextResponse', postTextResponse_httpStatus - The response's http status code.

PostContent

data PostContent

See: newPostContent smart constructor.

Constructors

PostContent' (Maybe Text) (Maybe (Sensitive Text)) (Maybe (Sensitive Text)) (Maybe (Sensitive Text)) Text Text Text Text HashedBody

newPostContent

Arguments

:: Text
$sel:botName:PostContent'

-> Text
$sel:botAlias:PostContent'

-> Text
$sel:userId:PostContent'

-> Text
$sel:contentType:PostContent'

-> HashedBody
$sel:inputStream:PostContent'

-> PostContent

Create a value of PostContent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accept:PostContent', postContent_accept - You pass this value as the Accept HTTP header.

The message Amazon Lex returns in the response can be either text or speech based on the Accept HTTP header value in the request.

If the value is text/plain; charset=utf-8, Amazon Lex returns text in the response.

If the value begins with audio/, Amazon Lex returns speech in the response. Amazon Lex uses Amazon Polly to generate the speech (using the configuration you specified in the Accept header). For example, if you specify audio/mpeg as the value, Amazon Lex returns speech in the MPEG format.

If the value is audio/pcm, the speech returned is audio/pcm in 16-bit, little endian format.

The following are the accepted values:

audio/mpeg

audio/ogg

audio/pcm

text/plain; charset=utf-8

audio/* (defaults to mpeg)

$sel:activeContexts:PostContent', postContent_activeContexts - A list of contexts active for the request. A context can be activated when a previous intent is fulfilled, or by including the context in the request,

If you don't specify a list of contexts, Amazon Lex will use the current list of contexts for the session. If you specify an empty list, all contexts for the session are cleared.

$sel:requestAttributes:PostContent', postContent_requestAttributes - You pass this value as the x-amz-lex-request-attributes HTTP header.

Request-specific information passed between Amazon Lex and a client application. The value must be a JSON serialized and base64 encoded map with string keys and values. The total size of the requestAttributes and sessionAttributes headers is limited to 12 KB.

The namespace x-amz-lex: is reserved for special attributes. Don't create any request attributes with the prefix x-amz-lex:.

For more information, see Setting Request Attributes.

$sel:sessionAttributes:PostContent', postContent_sessionAttributes - You pass this value as the x-amz-lex-session-attributes HTTP header.

Application-specific information passed between Amazon Lex and a client application. The value must be a JSON serialized and base64 encoded map with string keys and values. The total size of the sessionAttributes and requestAttributes headers is limited to 12 KB.

For more information, see Setting Session Attributes.

$sel:botName:PostContent', postContent_botName - Name of the Amazon Lex bot.

$sel:botAlias:PostContent', postContent_botAlias - Alias of the Amazon Lex bot.

$sel:userId:PostContent', postContent_userId - The ID of the client application user. Amazon Lex uses this to identify a user's conversation with your bot. At runtime, each request must contain the userID field.

To decide the user ID to use for your application, consider the following factors.

The userID field must not contain any personally identifiable information of the user, for example, name, personal identification numbers, or other end user personal information.

If you want a user to start a conversation on one device and continue on another device, use a user-specific identifier.

If you want the same user to be able to have two independent conversations on two different devices, choose a device-specific identifier.

A user can't have two independent conversations with two different versions of the same bot. For example, a user can't have a conversation with the PROD and BETA versions of the same bot. If you anticipate that a user will need to have conversation with two different versions, for example, while testing, include the bot alias in the user ID to separate the two conversations.

$sel:contentType:PostContent', postContent_contentType - You pass this value as the Content-Type HTTP header.

Indicates the audio format or text. The header value must start with one of the following prefixes:

PCM format, audio data must be in little-endian byte order.

audio/l16; rate=16000; channels=1

audio/x-l16; sample-rate=16000; channel-count=1

audio/lpcm; sample-rate=8000; sample-size-bits=16; channel-count=1; is-big-endian=false

Opus format

audio/x-cbr-opus-with-preamble; preamble-size=0; bit-rate=256000; frame-size-milliseconds=4

Text format

text/plain; charset=utf-8

$sel:inputStream:PostContent', postContent_inputStream - User input in PCM or Opus audio format or text format as described in the Content-Type HTTP header.

You can stream audio data to Amazon Lex or you can create a local buffer that captures all of the audio data before sending. In general, you get better performance if you stream audio data rather than buffering the data locally.

data PostContentResponse

See: newPostContentResponse smart constructor.

Constructors

PostContentResponse' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe (Sensitive Text)) (Maybe Text) (Maybe Text) (Maybe (Sensitive Text)) (Maybe DialogState) (Maybe (Sensitive Text)) (Maybe Text) (Maybe Text) (Maybe MessageFormatType) (Maybe (Sensitive Text)) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) Int ResponseBody

newPostContentResponse

Arguments

:: Int
$sel:httpStatus:PostContentResponse'

-> ResponseBody
$sel:audioStream:PostContentResponse'

-> PostContentResponse

Create a value of PostContentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sentimentResponse:PostContentResponse', postContentResponse_sentimentResponse - The sentiment expressed in an utterance.

When the bot is configured to send utterances to Amazon Comprehend for sentiment analysis, this field contains the result of the analysis.

$sel:nluIntentConfidence:PostContentResponse', postContentResponse_nluIntentConfidence - Provides a score that indicates how confident Amazon Lex is that the returned intent is the one that matches the user's intent. The score is between 0.0 and 1.0.

The score is a relative score, not an absolute score. The score may change based on improvements to Amazon Lex.

$sel:slots:PostContentResponse', postContentResponse_slots - Map of zero or more intent slots (name/value pairs) Amazon Lex detected from the user input during the conversation. The field is base-64 encoded.

Amazon Lex creates a resolution list containing likely values for a slot. The value that it returns is determined by the valueSelectionStrategy selected when the slot type was created or updated. If valueSelectionStrategy is set to ORIGINAL_VALUE, the value provided by the user is returned, if the user value is similar to the slot values. If valueSelectionStrategy is set to TOP_RESOLUTION Amazon Lex returns the first value in the resolution list or, if there is no resolution list, null. If you don't specify a valueSelectionStrategy, the default is ORIGINAL_VALUE.

$sel:encodedInputTranscript:PostContentResponse', postContentResponse_encodedInputTranscript - The text used to process the request.

If the input was an audio stream, the encodedInputTranscript field contains the text extracted from the audio stream. This is the text that is actually processed to recognize intents and slot values. You can use this information to determine if Amazon Lex is correctly processing the audio that you send.

The encodedInputTranscript field is base-64 encoded. You must decode the field before you can use the value.

$sel:intentName:PostContentResponse', postContentResponse_intentName - Current user intent that Amazon Lex is aware of.

$sel:botVersion:PostContentResponse', postContentResponse_botVersion - The version of the bot that responded to the conversation. You can use this information to help determine if one version of a bot is performing better than another version.

$sel:encodedMessage:PostContentResponse', postContentResponse_encodedMessage - The message to convey to the user. The message can come from the bot's configuration or from a Lambda function.

If the intent is not configured with a Lambda function, or if the Lambda function returned Delegate as the dialogAction.type in its response, Amazon Lex decides on the next course of action and selects an appropriate message from the bot's configuration based on the current interaction context. For example, if Amazon Lex isn't able to understand user input, it uses a clarification prompt message.

When you create an intent you can assign messages to groups. When messages are assigned to groups Amazon Lex returns one message from each group in the response. The message field is an escaped JSON string containing the messages. For more information about the structure of the JSON string returned, see msg-prompts-formats.

If the Lambda function returns a message, Amazon Lex passes it to the client in its response.

The encodedMessage field is base-64 encoded. You must decode the field before you can use the value.

$sel:dialogState:PostContentResponse', postContentResponse_dialogState - Identifies the current state of the user interaction. Amazon Lex returns one of the following values as dialogState. The client can optionally use this information to customize the user interface.

ElicitIntent - Amazon Lex wants to elicit the user's intent. Consider the following examples:

For example, a user might utter an intent ("I want to order a pizza"). If Amazon Lex cannot infer the user intent from this utterance, it will return this dialog state.

ConfirmIntent - Amazon Lex is expecting a "yes" or "no" response.

For example, Amazon Lex wants user confirmation before fulfilling an intent. Instead of a simple "yes" or "no" response, a user might respond with additional information. For example, "yes, but make it a thick crust pizza" or "no, I want to order a drink." Amazon Lex can process such additional information (in these examples, update the crust type slot or change the intent from OrderPizza to OrderDrink).

ElicitSlot - Amazon Lex is expecting the value of a slot for the current intent.

For example, suppose that in the response Amazon Lex sends this message: "What size pizza would you like?". A user might reply with the slot value (e.g., "medium"). The user might also provide additional information in the response (e.g., "medium thick crust pizza"). Amazon Lex can process such additional information appropriately.

Fulfilled - Conveys that the Lambda function has successfully fulfilled the intent.

ReadyForFulfillment - Conveys that the client has to fulfill the request.

Failed - Conveys that the conversation with the user failed.

This can happen for various reasons, including that the user does not provide an appropriate response to prompts from the service (you can configure how many times Amazon Lex can prompt a user for specific information), or if the Lambda function fails to fulfill the intent.

$sel:activeContexts:PostContent', postContentResponse_activeContexts - A list of active contexts for the session. A context can be set when an intent is fulfilled or by calling the PostContent, PostText, or PutSession operation.

You can use a context to control the intents that can follow up an intent, or to modify the operation of your application.

$sel:alternativeIntents:PostContentResponse', postContentResponse_alternativeIntents - One to four alternative intents that may be applicable to the user's intent.

Each alternative includes a score that indicates how confident Amazon Lex is that the intent matches the user's intent. The intents are sorted by the confidence score.

$sel:inputTranscript:PostContentResponse', postContentResponse_inputTranscript - The text used to process the request.

You can use this field only in the de-DE, en-AU, en-GB, en-US, es-419, es-ES, es-US, fr-CA, fr-FR, and it-IT locales. In all other locales, the inputTranscript field is null. You should use the encodedInputTranscript field instead.

If the input was an audio stream, the inputTranscript field contains the text extracted from the audio stream. This is the text that is actually processed to recognize intents and slot values. You can use this information to determine if Amazon Lex is correctly processing the audio that you send.

$sel:messageFormat:PostContentResponse', postContentResponse_messageFormat - The format of the response message. One of the following values:

PlainText - The message contains plain UTF-8 text.

CustomPayload - The message is a custom format for the client.

SSML - The message contains text formatted for voice output.

Composite - The message contains an escaped JSON object containing one or more messages from the groups that messages were assigned to when the intent was created.

$sel:message:PostContentResponse', postContentResponse_message - You can only use this field in the de-DE, en-AU, en-GB, en-US, es-419, es-ES, es-US, fr-CA, fr-FR, and it-IT locales. In all other locales, the message field is null. You should use the encodedMessage field instead.

The message to convey to the user. The message can come from the bot's configuration or from a Lambda function.

If the intent is not configured with a Lambda function, or if the Lambda function returned Delegate as the dialogAction.type in its response, Amazon Lex decides on the next course of action and selects an appropriate message from the bot's configuration based on the current interaction context. For example, if Amazon Lex isn't able to understand user input, it uses a clarification prompt message.

When you create an intent you can assign messages to groups. When messages are assigned to groups Amazon Lex returns one message from each group in the response. The message field is an escaped JSON string containing the messages. For more information about the structure of the JSON string returned, see msg-prompts-formats.

If the Lambda function returns a message, Amazon Lex passes it to the client in its response.

$sel:sessionId:PostContentResponse', postContentResponse_sessionId - The unique identifier for the session.

$sel:slotToElicit:PostContentResponse', postContentResponse_slotToElicit - If the dialogState value is ElicitSlot, returns the name of the slot for which Amazon Lex is eliciting a value.

$sel:contentType:PostContent', postContentResponse_contentType - Content type as specified in the Accept HTTP header in the request.

$sel:sessionAttributes:PostContent', postContentResponse_sessionAttributes - Map of key/value pairs representing the session-specific context information.

$sel:httpStatus:PostContentResponse', postContentResponse_httpStatus - The response's http status code.

$sel:audioStream:PostContentResponse', postContentResponse_audioStream - The prompt (or statement) to convey to the user. This is based on the bot configuration and context. For example, if Amazon Lex did not understand the user intent, it sends the clarificationPrompt configured for the bot. If the intent requires confirmation before taking the fulfillment action, it sends the confirmationPrompt. Another example: Suppose that the Lambda function successfully fulfilled the intent, and sent a message to convey to the user. Then Amazon Lex sends that message in the response.

GetSession

data GetSession

See: newGetSession smart constructor.

Constructors

GetSession' (Maybe Text) Text Text Text

newGetSession

Arguments

:: Text
$sel:botName:GetSession'

-> Text
$sel:botAlias:GetSession'

-> Text
$sel:userId:GetSession'

-> GetSession

Create a value of GetSession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checkpointLabelFilter:GetSession', getSession_checkpointLabelFilter - A string used to filter the intents returned in the recentIntentSummaryView structure.

When you specify a filter, only intents with their checkpointLabel field set to that string are returned.

$sel:botName:GetSession', getSession_botName - The name of the bot that contains the session data.

$sel:botAlias:GetSession', getSession_botAlias - The alias in use for the bot that contains the session data.

$sel:userId:GetSession', getSession_userId - The ID of the client application user. Amazon Lex uses this to identify a user's conversation with your bot.

data GetSessionResponse

See: newGetSessionResponse smart constructor.

Constructors

GetSessionResponse' (Maybe (Sensitive [ActiveContext])) (Maybe Text) (Maybe [IntentSummary]) (Maybe DialogAction) (Maybe (Sensitive (HashMap Text Text))) Int

newGetSessionResponse

Arguments

:: Int
$sel:httpStatus:GetSessionResponse'

-> GetSessionResponse

Create a value of GetSessionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:activeContexts:GetSessionResponse', getSessionResponse_activeContexts - A list of active contexts for the session. A context can be set when an intent is fulfilled or by calling the PostContent, PostText, or PutSession operation.

You can use a context to control the intents that can follow up an intent, or to modify the operation of your application.

$sel:sessionId:GetSessionResponse', getSessionResponse_sessionId - A unique identifier for the session.

$sel:recentIntentSummaryView:GetSessionResponse', getSessionResponse_recentIntentSummaryView - An array of information about the intents used in the session. The array can contain a maximum of three summaries. If more than three intents are used in the session, the recentIntentSummaryView operation contains information about the last three intents used.

If you set the checkpointLabelFilter parameter in the request, the array contains only the intents with the specified label.

$sel:dialogAction:GetSessionResponse', getSessionResponse_dialogAction - Describes the current state of the bot.

$sel:sessionAttributes:GetSessionResponse', getSessionResponse_sessionAttributes - Map of key/value pairs representing the session-specific context information. It contains application information passed between Amazon Lex and a client application.

$sel:httpStatus:GetSessionResponse', getSessionResponse_httpStatus - The response's http status code.

Types

ConfirmationStatus

newtype ConfirmationStatus

Constructors

ConfirmationStatus'

Fields

fromConfirmationStatus :: Text

Bundled Patterns

pattern ConfirmationStatus_Confirmed :: ConfirmationStatus

pattern ConfirmationStatus_Denied :: ConfirmationStatus

pattern ConfirmationStatus_None :: ConfirmationStatus

ContentType

newtype ContentType

Constructors

ContentType'

Fields

fromContentType :: Text

Bundled Patterns

pattern ContentType_Application_vnd_amazonaws_card_generic :: ContentType

DialogActionType

newtype DialogActionType

Constructors

DialogActionType'

Fields

fromDialogActionType :: Text

Bundled Patterns

pattern DialogActionType_Close :: DialogActionType

pattern DialogActionType_ConfirmIntent :: DialogActionType

pattern DialogActionType_Delegate :: DialogActionType

pattern DialogActionType_ElicitIntent :: DialogActionType

pattern DialogActionType_ElicitSlot :: DialogActionType

DialogState

newtype DialogState

Constructors

DialogState'

Fields

fromDialogState :: Text

Bundled Patterns

pattern DialogState_ConfirmIntent :: DialogState

pattern DialogState_ElicitIntent :: DialogState

pattern DialogState_ElicitSlot :: DialogState

pattern DialogState_Failed :: DialogState

pattern DialogState_Fulfilled :: DialogState

pattern DialogState_ReadyForFulfillment :: DialogState

FulfillmentState

newtype FulfillmentState

Constructors

FulfillmentState'

Fields

fromFulfillmentState :: Text

Bundled Patterns

pattern FulfillmentState_Failed :: FulfillmentState

pattern FulfillmentState_Fulfilled :: FulfillmentState

pattern FulfillmentState_ReadyForFulfillment :: FulfillmentState

MessageFormatType

newtype MessageFormatType

Constructors

MessageFormatType'

Fields

fromMessageFormatType :: Text

Bundled Patterns

pattern MessageFormatType_Composite :: MessageFormatType

pattern MessageFormatType_CustomPayload :: MessageFormatType

pattern MessageFormatType_PlainText :: MessageFormatType

pattern MessageFormatType_SSML :: MessageFormatType

ActiveContext

data ActiveContext

A context is a variable that contains information about the current state of the conversation between a user and Amazon Lex. Context can be set automatically by Amazon Lex when an intent is fulfilled, or it can be set at runtime using the PutContent, PutText, or PutSession operation.

See: newActiveContext smart constructor.

Constructors

ActiveContext' Text ActiveContextTimeToLive (HashMap Text (Sensitive Text))

newActiveContext

Arguments

:: Text
$sel:name:ActiveContext'

-> ActiveContextTimeToLive
$sel:timeToLive:ActiveContext'

-> ActiveContext

Create a value of ActiveContext with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ActiveContext', activeContext_name - The name of the context.

$sel:timeToLive:ActiveContext', activeContext_timeToLive - The length of time or number of turns that a context remains active.

$sel:parameters:ActiveContext', activeContext_parameters - State variables for the current context. You can use these values as default values for slots in subsequent events.

ActiveContextTimeToLive

data ActiveContextTimeToLive

The length of time or number of turns that a context remains active.

See: newActiveContextTimeToLive smart constructor.

Constructors

ActiveContextTimeToLive' (Maybe Natural) (Maybe Natural)

newActiveContextTimeToLive :: ActiveContextTimeToLive

Create a value of ActiveContextTimeToLive with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:turnsToLive:ActiveContextTimeToLive', activeContextTimeToLive_turnsToLive - The number of conversation turns that the context should be active. A conversation turn is one PostContent or PostText request and the corresponding response from Amazon Lex.

$sel:timeToLiveInSeconds:ActiveContextTimeToLive', activeContextTimeToLive_timeToLiveInSeconds - The number of seconds that the context should be active after it is first sent in a PostContent or PostText response. You can set the value between 5 and 86,400 seconds (24 hours).

Button

data Button

Represents an option to be shown on the client platform (Facebook, Slack, etc.)

See: newButton smart constructor.

Constructors

Button' Text Text

newButton

Arguments

:: Text
$sel:text:Button'

-> Text
$sel:value:Button'

-> Button

Create a value of Button with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:text:Button', button_text - Text that is visible to the user on the button.

$sel:value:Button', button_value - The value sent to Amazon Lex when a user chooses the button. For example, consider button text "NYC." When the user chooses the button, the value sent can be "New York City."

DialogAction

data DialogAction

Describes the next action that the bot should take in its interaction with the user and provides information about the context in which the action takes place. Use the DialogAction data type to set the interaction to a specific state, or to return the interaction to a previous state.

See: newDialogAction smart constructor.

Constructors

DialogAction' (Maybe (Sensitive (HashMap Text Text))) (Maybe Text) (Maybe FulfillmentState) (Maybe MessageFormatType) (Maybe (Sensitive Text)) (Maybe Text) DialogActionType

newDialogAction

Arguments

:: DialogActionType
$sel:type':DialogAction'

-> DialogAction

Create a value of DialogAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:slots:DialogAction', dialogAction_slots - Map of the slots that have been gathered and their values.

$sel:intentName:DialogAction', dialogAction_intentName - The name of the intent.

$sel:fulfillmentState:DialogAction', dialogAction_fulfillmentState - The fulfillment state of the intent. The possible values are:

Failed - The Lambda function associated with the intent failed to fulfill the intent.

Fulfilled - The intent has fulfilled by the Lambda function associated with the intent.

ReadyForFulfillment - All of the information necessary for the intent is present and the intent ready to be fulfilled by the client application.

$sel:messageFormat:DialogAction', dialogAction_messageFormat - - PlainText - The message contains plain UTF-8 text.

CustomPayload - The message is a custom format for the client.

SSML - The message contains text formatted for voice output.

Composite - The message contains an escaped JSON object containing one or more messages. For more information, see Message Groups.

$sel:message:DialogAction', dialogAction_message - The message that should be shown to the user. If you don't specify a message, Amazon Lex will use the message configured for the intent.

$sel:slotToElicit:DialogAction', dialogAction_slotToElicit - The name of the slot that should be elicited from the user.

$sel:type':DialogAction', dialogAction_type - The next action that the bot should take in its interaction with the user. The possible values are:

ConfirmIntent - The next action is asking the user if the intent is complete and ready to be fulfilled. This is a yes/no question such as "Place the order?"

Close - Indicates that the there will not be a response from the user. For example, the statement "Your order has been placed" does not require a response.

Delegate - The next action is determined by Amazon Lex.

ElicitIntent - The next action is to determine the intent that the user wants to fulfill.

ElicitSlot - The next action is to elicit a slot value from the user.

GenericAttachment

data GenericAttachment

Represents an option rendered to the user when a prompt is shown. It could be an image, a button, a link, or text.

See: newGenericAttachment smart constructor.

Constructors

GenericAttachment' (Maybe [Button]) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text)

newGenericAttachment :: GenericAttachment

Create a value of GenericAttachment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:buttons:GenericAttachment', genericAttachment_buttons - The list of options to show to the user.

$sel:subTitle:GenericAttachment', genericAttachment_subTitle - The subtitle shown below the title.

$sel:imageUrl:GenericAttachment', genericAttachment_imageUrl - The URL of an image that is displayed to the user.

$sel:attachmentLinkUrl:GenericAttachment', genericAttachment_attachmentLinkUrl - The URL of an attachment to the response card.

$sel:title:GenericAttachment', genericAttachment_title - The title of the option.

IntentConfidence

data IntentConfidence

Provides a score that indicates the confidence that Amazon Lex has that an intent is the one that satisfies the user's intent.

See: newIntentConfidence smart constructor.

Constructors

IntentConfidence' (Maybe Double)

newIntentConfidence :: IntentConfidence

Create a value of IntentConfidence with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:score:IntentConfidence', intentConfidence_score - A score that indicates how confident Amazon Lex is that an intent satisfies the user's intent. Ranges between 0.00 and 1.00. Higher scores indicate higher confidence.

IntentSummary

data IntentSummary

Provides information about the state of an intent. You can use this information to get the current state of an intent so that you can process the intent, or so that you can return the intent to its previous state.

See: newIntentSummary smart constructor.

Constructors

IntentSummary' (Maybe Text) (Maybe (Sensitive (HashMap Text Text))) (Maybe Text) (Maybe FulfillmentState) (Maybe ConfirmationStatus) (Maybe Text) DialogActionType

newIntentSummary

Arguments

:: DialogActionType
$sel:dialogActionType:IntentSummary'

-> IntentSummary

Create a value of IntentSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:checkpointLabel:IntentSummary', intentSummary_checkpointLabel - A user-defined label that identifies a particular intent. You can use this label to return to a previous intent.

Use the checkpointLabelFilter parameter of the GetSessionRequest operation to filter the intents returned by the operation to those with only the specified label.

$sel:slots:IntentSummary', intentSummary_slots - Map of the slots that have been gathered and their values.

$sel:intentName:IntentSummary', intentSummary_intentName - The name of the intent.

$sel:fulfillmentState:IntentSummary', intentSummary_fulfillmentState - The fulfillment state of the intent. The possible values are:

Failed - The Lambda function associated with the intent failed to fulfill the intent.

Fulfilled - The intent has fulfilled by the Lambda function associated with the intent.

ReadyForFulfillment - All of the information necessary for the intent is present and the intent ready to be fulfilled by the client application.

$sel:confirmationStatus:IntentSummary', intentSummary_confirmationStatus - The status of the intent after the user responds to the confirmation prompt. If the user confirms the intent, Amazon Lex sets this field to Confirmed. If the user denies the intent, Amazon Lex sets this value to Denied. The possible values are:

Confirmed - The user has responded "Yes" to the confirmation prompt, confirming that the intent is complete and that it is ready to be fulfilled.

Denied - The user has responded "No" to the confirmation prompt.

None - The user has never been prompted for confirmation; or, the user was prompted but did not confirm or deny the prompt.

$sel:slotToElicit:IntentSummary', intentSummary_slotToElicit - The next slot to elicit from the user. If there is not slot to elicit, the field is blank.

$sel:dialogActionType:IntentSummary', intentSummary_dialogActionType - The next action that the bot should take in its interaction with the user. The possible values are:

ConfirmIntent - The next action is asking the user if the intent is complete and ready to be fulfilled. This is a yes/no question such as "Place the order?"

Close - Indicates that the there will not be a response from the user. For example, the statement "Your order has been placed" does not require a response.

ElicitIntent - The next action is to determine the intent that the user wants to fulfill.

ElicitSlot - The next action is to elicit a slot value from the user.

PredictedIntent

data PredictedIntent

An intent that Amazon Lex suggests satisfies the user's intent. Includes the name of the intent, the confidence that Amazon Lex has that the user's intent is satisfied, and the slots defined for the intent.

See: newPredictedIntent smart constructor.

Constructors

PredictedIntent' (Maybe IntentConfidence) (Maybe (Sensitive (HashMap Text Text))) (Maybe Text)

newPredictedIntent :: PredictedIntent

Create a value of PredictedIntent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nluIntentConfidence:PredictedIntent', predictedIntent_nluIntentConfidence - Indicates how confident Amazon Lex is that an intent satisfies the user's intent.

$sel:slots:PredictedIntent', predictedIntent_slots - The slot and slot values associated with the predicted intent.

$sel:intentName:PredictedIntent', predictedIntent_intentName - The name of the intent that Amazon Lex suggests satisfies the user's intent.

ResponseCard

data ResponseCard

If you configure a response card when creating your bots, Amazon Lex substitutes the session attributes and slot values that are available, and then returns it. The response card can also come from a Lambda function ( dialogCodeHook and fulfillmentActivity on an intent).

See: newResponseCard smart constructor.

Constructors

ResponseCard' (Maybe [GenericAttachment]) (Maybe Text) (Maybe ContentType)

newResponseCard :: ResponseCard

Create a value of ResponseCard with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:genericAttachments:ResponseCard', responseCard_genericAttachments - An array of attachment objects representing options.

$sel:version:ResponseCard', responseCard_version - The version of the response card format.

$sel:contentType:ResponseCard', responseCard_contentType - The content type of the response.

SentimentResponse

data SentimentResponse

The sentiment expressed in an utterance.

When the bot is configured to send utterances to Amazon Comprehend for sentiment analysis, this field structure contains the result of the analysis.

See: newSentimentResponse smart constructor.

Constructors

SentimentResponse' (Maybe Text) (Maybe Text)

newSentimentResponse :: SentimentResponse

Create a value of SentimentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sentimentScore:SentimentResponse', sentimentResponse_sentimentScore - The likelihood that the sentiment was correctly inferred.

$sel:sentimentLabel:SentimentResponse', sentimentResponse_sentimentLabel - The inferred sentiment that Amazon Comprehend has the highest confidence in.