Network.AWS.LexRuntime.PostText

Description

Sends user input to Amazon Lex. Client applications can use this API to send requests to Amazon Lex at runtime. Amazon Lex then interprets the user input using the machine learning model it built for the bot.

In response, Amazon Lex returns the next message to convey to the user an optional responseCard to display. Consider the following example messages:

For a user input "I would like a pizza", Amazon Lex might return a response with a message eliciting slot data (for example, PizzaSize): "What size pizza would you like?"

After the user provides all of the pizza order information, Amazon Lex might return a response with a message to obtain user confirmation "Proceed with the pizza order?".

After the user replies to a confirmation prompt with a "yes", Amazon Lex might return a conclusion statement: "Thank you, your cheese pizza has been ordered.".

Not all Amazon Lex messages require a user response. For example, a conclusion statement does not require a response. Some messages require only a "yes" or "no" user response. In addition to the message, Amazon Lex provides additional context about the message in the response that you might use to enhance client behavior, for example, to display the appropriate client user interface. These are the slotToElicit, dialogState, intentName, and slots fields in the response. Consider the following examples:

If the message is to elicit slot data, Amazon Lex returns the following context information:

dialogState set to ElicitSlot

intentName set to the intent name in the current context

slotToElicit set to the slot name for which the message is eliciting information

slots set to a map of slots, configured for the intent, with currently known values

If the message is a confirmation prompt, the dialogState is set to ConfirmIntent and SlotToElicit is set to null.

If the message is a clarification prompt (configured for the intent) that indicates that user intent is not understood, the dialogState is set to ElicitIntent and slotToElicit is set to null.

In addition, Amazon Lex also returns your application-specific sessionAttributes. For more information, see Managing Conversation Context.

Creating a Request

data PostText

See: newPostText smart constructor.

Constructors

PostText'

Fields

activeContexts :: Maybe (Sensitive [ActiveContext])

A list of contexts active for the request. A context can be activated when a previous intent is fulfilled, or by including the context in the request,

If you don't specify a list of contexts, Amazon Lex will use the current list of contexts for the session. If you specify an empty list, all contexts for the session are cleared.

requestAttributes :: Maybe (Sensitive (HashMap Text Text))

Request-specific information passed between Amazon Lex and a client application.

The namespace x-amz-lex: is reserved for special attributes. Don't create any request attributes with the prefix x-amz-lex:.

For more information, see Setting Request Attributes.

sessionAttributes :: Maybe (Sensitive (HashMap Text Text))

Application-specific information passed between Amazon Lex and a client application.

For more information, see Setting Session Attributes.

botName :: Text

The name of the Amazon Lex bot.

botAlias :: Text

The alias of the Amazon Lex bot.

userId :: Text

The ID of the client application user. Amazon Lex uses this to identify a user's conversation with your bot. At runtime, each request must contain the userID field.

To decide the user ID to use for your application, consider the following factors.

The userID field must not contain any personally identifiable information of the user, for example, name, personal identification numbers, or other end user personal information.

If you want a user to start a conversation on one device and continue on another device, use a user-specific identifier.

If you want the same user to be able to have two independent conversations on two different devices, choose a device-specific identifier.

A user can't have two independent conversations with two different versions of the same bot. For example, a user can't have a conversation with the PROD and BETA versions of the same bot. If you anticipate that a user will need to have conversation with two different versions, for example, while testing, include the bot alias in the user ID to separate the two conversations.

inputText :: Sensitive Text

The text that the user entered (Amazon Lex interprets this text).

newPostText

Arguments

:: Text
$sel:botName:PostText'

-> Text
$sel:botAlias:PostText'

-> Text
$sel:userId:PostText'

-> Text
$sel:inputText:PostText'

-> PostText

Create a value of PostText with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:activeContexts:PostText', postText_activeContexts - A list of contexts active for the request. A context can be activated when a previous intent is fulfilled, or by including the context in the request,

If you don't specify a list of contexts, Amazon Lex will use the current list of contexts for the session. If you specify an empty list, all contexts for the session are cleared.

$sel:requestAttributes:PostText', postText_requestAttributes - Request-specific information passed between Amazon Lex and a client application.

The namespace x-amz-lex: is reserved for special attributes. Don't create any request attributes with the prefix x-amz-lex:.

For more information, see Setting Request Attributes.

$sel:sessionAttributes:PostText', postText_sessionAttributes - Application-specific information passed between Amazon Lex and a client application.

For more information, see Setting Session Attributes.

$sel:botName:PostText', postText_botName - The name of the Amazon Lex bot.

$sel:botAlias:PostText', postText_botAlias - The alias of the Amazon Lex bot.

$sel:userId:PostText', postText_userId - The ID of the client application user. Amazon Lex uses this to identify a user's conversation with your bot. At runtime, each request must contain the userID field.

To decide the user ID to use for your application, consider the following factors.

The userID field must not contain any personally identifiable information of the user, for example, name, personal identification numbers, or other end user personal information.

If you want a user to start a conversation on one device and continue on another device, use a user-specific identifier.

If you want the same user to be able to have two independent conversations on two different devices, choose a device-specific identifier.

A user can't have two independent conversations with two different versions of the same bot. For example, a user can't have a conversation with the PROD and BETA versions of the same bot. If you anticipate that a user will need to have conversation with two different versions, for example, while testing, include the bot alias in the user ID to separate the two conversations.

$sel:inputText:PostText', postText_inputText - The text that the user entered (Amazon Lex interprets this text).

Request Lenses

postText_activeContexts :: Lens' PostText (Maybe [ActiveContext])

A list of contexts active for the request. A context can be activated when a previous intent is fulfilled, or by including the context in the request,

If you don't specify a list of contexts, Amazon Lex will use the current list of contexts for the session. If you specify an empty list, all contexts for the session are cleared.

postText_requestAttributes :: Lens' PostText (Maybe (HashMap Text Text))

Request-specific information passed between Amazon Lex and a client application.

The namespace x-amz-lex: is reserved for special attributes. Don't create any request attributes with the prefix x-amz-lex:.

For more information, see Setting Request Attributes.

postText_sessionAttributes :: Lens' PostText (Maybe (HashMap Text Text))

Application-specific information passed between Amazon Lex and a client application.

For more information, see Setting Session Attributes.

postText_botName :: Lens' PostText Text

The name of the Amazon Lex bot.

postText_botAlias :: Lens' PostText Text

The alias of the Amazon Lex bot.

postText_userId :: Lens' PostText Text

The ID of the client application user. Amazon Lex uses this to identify a user's conversation with your bot. At runtime, each request must contain the userID field.

To decide the user ID to use for your application, consider the following factors.

The userID field must not contain any personally identifiable information of the user, for example, name, personal identification numbers, or other end user personal information.

If you want a user to start a conversation on one device and continue on another device, use a user-specific identifier.

If you want the same user to be able to have two independent conversations on two different devices, choose a device-specific identifier.

A user can't have two independent conversations with two different versions of the same bot. For example, a user can't have a conversation with the PROD and BETA versions of the same bot. If you anticipate that a user will need to have conversation with two different versions, for example, while testing, include the bot alias in the user ID to separate the two conversations.

postText_inputText :: Lens' PostText Text

The text that the user entered (Amazon Lex interprets this text).

Destructuring the Response

data PostTextResponse

See: newPostTextResponse smart constructor.

Constructors

PostTextResponse'

Fields

sentimentResponse :: Maybe SentimentResponse

The sentiment expressed in and utterance.

When the bot is configured to send utterances to Amazon Comprehend for sentiment analysis, this field contains the result of the analysis.

nluIntentConfidence :: Maybe IntentConfidence

Provides a score that indicates how confident Amazon Lex is that the returned intent is the one that matches the user's intent. The score is between 0.0 and 1.0. For more information, see Confidence Scores.

The score is a relative score, not an absolute score. The score may change based on improvements to Amazon Lex.

slots :: Maybe (Sensitive (HashMap Text Text))

The intent slots that Amazon Lex detected from the user input in the conversation.

Amazon Lex creates a resolution list containing likely values for a slot. The value that it returns is determined by the valueSelectionStrategy selected when the slot type was created or updated. If valueSelectionStrategy is set to ORIGINAL_VALUE, the value provided by the user is returned, if the user value is similar to the slot values. If valueSelectionStrategy is set to TOP_RESOLUTION Amazon Lex returns the first value in the resolution list or, if there is no resolution list, null. If you don't specify a valueSelectionStrategy, the default is ORIGINAL_VALUE.

responseCard :: Maybe ResponseCard

Represents the options that the user has to respond to the current prompt. Response Card can come from the bot configuration (in the Amazon Lex console, choose the settings button next to a slot) or from a code hook (Lambda function).

intentName :: Maybe Text

The current user intent that Amazon Lex is aware of.

botVersion :: Maybe Text

The version of the bot that responded to the conversation. You can use this information to help determine if one version of a bot is performing better than another version.

dialogState :: Maybe DialogState

Identifies the current state of the user interaction. Amazon Lex returns one of the following values as dialogState. The client can optionally use this information to customize the user interface.

ElicitIntent - Amazon Lex wants to elicit user intent.

For example, a user might utter an intent ("I want to order a pizza"). If Amazon Lex cannot infer the user intent from this utterance, it will return this dialogState.

ConfirmIntent - Amazon Lex is expecting a "yes" or "no" response.

For example, Amazon Lex wants user confirmation before fulfilling an intent.

Instead of a simple "yes" or "no," a user might respond with additional information. For example, "yes, but make it thick crust pizza" or "no, I want to order a drink". Amazon Lex can process such additional information (in these examples, update the crust type slot value, or change intent from OrderPizza to OrderDrink).

ElicitSlot - Amazon Lex is expecting a slot value for the current intent.

For example, suppose that in the response Amazon Lex sends this message: "What size pizza would you like?". A user might reply with the slot value (e.g., "medium"). The user might also provide additional information in the response (e.g., "medium thick crust pizza"). Amazon Lex can process such additional information appropriately.

Fulfilled - Conveys that the Lambda function configured for the intent has successfully fulfilled the intent.

ReadyForFulfillment - Conveys that the client has to fulfill the intent.

Failed - Conveys that the conversation with the user failed.

This can happen for various reasons including that the user did not provide an appropriate response to prompts from the service (you can configure how many times Amazon Lex can prompt a user for specific information), or the Lambda function failed to fulfill the intent.

activeContexts :: Maybe (Sensitive [ActiveContext])

A list of active contexts for the session. A context can be set when an intent is fulfilled or by calling the PostContent, PostText, or PutSession operation.

You can use a context to control the intents that can follow up an intent, or to modify the operation of your application.

alternativeIntents :: Maybe [PredictedIntent]

One to four alternative intents that may be applicable to the user's intent.

Each alternative includes a score that indicates how confident Amazon Lex is that the intent matches the user's intent. The intents are sorted by the confidence score.

messageFormat :: Maybe MessageFormatType

The format of the response message. One of the following values:

PlainText - The message contains plain UTF-8 text.

CustomPayload - The message is a custom format defined by the Lambda function.

SSML - The message contains text formatted for voice output.

Composite - The message contains an escaped JSON object containing one or more messages from the groups that messages were assigned to when the intent was created.

message :: Maybe (Sensitive Text)

The message to convey to the user. The message can come from the bot's configuration or from a Lambda function.

If the intent is not configured with a Lambda function, or if the Lambda function returned Delegate as the dialogAction.type its response, Amazon Lex decides on the next course of action and selects an appropriate message from the bot's configuration based on the current interaction context. For example, if Amazon Lex isn't able to understand user input, it uses a clarification prompt message.

When you create an intent you can assign messages to groups. When messages are assigned to groups Amazon Lex returns one message from each group in the response. The message field is an escaped JSON string containing the messages. For more information about the structure of the JSON string returned, see msg-prompts-formats.

If the Lambda function returns a message, Amazon Lex passes it to the client in its response.

sessionId :: Maybe Text

A unique identifier for the session.

slotToElicit :: Maybe Text

If the dialogState value is ElicitSlot, returns the name of the slot for which Amazon Lex is eliciting a value.

sessionAttributes :: Maybe (Sensitive (HashMap Text Text))

A map of key-value pairs representing the session-specific context information.

httpStatus :: Int

The response's http status code.

newPostTextResponse

Arguments

:: Int
$sel:httpStatus:PostTextResponse'

-> PostTextResponse

Create a value of PostTextResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sentimentResponse:PostTextResponse', postTextResponse_sentimentResponse - The sentiment expressed in and utterance.

When the bot is configured to send utterances to Amazon Comprehend for sentiment analysis, this field contains the result of the analysis.

$sel:nluIntentConfidence:PostTextResponse', postTextResponse_nluIntentConfidence - Provides a score that indicates how confident Amazon Lex is that the returned intent is the one that matches the user's intent. The score is between 0.0 and 1.0. For more information, see Confidence Scores.

The score is a relative score, not an absolute score. The score may change based on improvements to Amazon Lex.

$sel:slots:PostTextResponse', postTextResponse_slots - The intent slots that Amazon Lex detected from the user input in the conversation.

Amazon Lex creates a resolution list containing likely values for a slot. The value that it returns is determined by the valueSelectionStrategy selected when the slot type was created or updated. If valueSelectionStrategy is set to ORIGINAL_VALUE, the value provided by the user is returned, if the user value is similar to the slot values. If valueSelectionStrategy is set to TOP_RESOLUTION Amazon Lex returns the first value in the resolution list or, if there is no resolution list, null. If you don't specify a valueSelectionStrategy, the default is ORIGINAL_VALUE.

$sel:responseCard:PostTextResponse', postTextResponse_responseCard - Represents the options that the user has to respond to the current prompt. Response Card can come from the bot configuration (in the Amazon Lex console, choose the settings button next to a slot) or from a code hook (Lambda function).

$sel:intentName:PostTextResponse', postTextResponse_intentName - The current user intent that Amazon Lex is aware of.

$sel:botVersion:PostTextResponse', postTextResponse_botVersion - The version of the bot that responded to the conversation. You can use this information to help determine if one version of a bot is performing better than another version.

$sel:dialogState:PostTextResponse', postTextResponse_dialogState - Identifies the current state of the user interaction. Amazon Lex returns one of the following values as dialogState. The client can optionally use this information to customize the user interface.

ElicitIntent - Amazon Lex wants to elicit user intent.

For example, a user might utter an intent ("I want to order a pizza"). If Amazon Lex cannot infer the user intent from this utterance, it will return this dialogState.

ConfirmIntent - Amazon Lex is expecting a "yes" or "no" response.

For example, Amazon Lex wants user confirmation before fulfilling an intent.

Instead of a simple "yes" or "no," a user might respond with additional information. For example, "yes, but make it thick crust pizza" or "no, I want to order a drink". Amazon Lex can process such additional information (in these examples, update the crust type slot value, or change intent from OrderPizza to OrderDrink).

ElicitSlot - Amazon Lex is expecting a slot value for the current intent.

For example, suppose that in the response Amazon Lex sends this message: "What size pizza would you like?". A user might reply with the slot value (e.g., "medium"). The user might also provide additional information in the response (e.g., "medium thick crust pizza"). Amazon Lex can process such additional information appropriately.

Fulfilled - Conveys that the Lambda function configured for the intent has successfully fulfilled the intent.

ReadyForFulfillment - Conveys that the client has to fulfill the intent.

Failed - Conveys that the conversation with the user failed.

This can happen for various reasons including that the user did not provide an appropriate response to prompts from the service (you can configure how many times Amazon Lex can prompt a user for specific information), or the Lambda function failed to fulfill the intent.

$sel:activeContexts:PostText', postTextResponse_activeContexts - A list of active contexts for the session. A context can be set when an intent is fulfilled or by calling the PostContent, PostText, or PutSession operation.

You can use a context to control the intents that can follow up an intent, or to modify the operation of your application.

$sel:alternativeIntents:PostTextResponse', postTextResponse_alternativeIntents - One to four alternative intents that may be applicable to the user's intent.

Each alternative includes a score that indicates how confident Amazon Lex is that the intent matches the user's intent. The intents are sorted by the confidence score.

$sel:messageFormat:PostTextResponse', postTextResponse_messageFormat - The format of the response message. One of the following values:

PlainText - The message contains plain UTF-8 text.

CustomPayload - The message is a custom format defined by the Lambda function.

SSML - The message contains text formatted for voice output.

Composite - The message contains an escaped JSON object containing one or more messages from the groups that messages were assigned to when the intent was created.

$sel:message:PostTextResponse', postTextResponse_message - The message to convey to the user. The message can come from the bot's configuration or from a Lambda function.

If the intent is not configured with a Lambda function, or if the Lambda function returned Delegate as the dialogAction.type its response, Amazon Lex decides on the next course of action and selects an appropriate message from the bot's configuration based on the current interaction context. For example, if Amazon Lex isn't able to understand user input, it uses a clarification prompt message.

When you create an intent you can assign messages to groups. When messages are assigned to groups Amazon Lex returns one message from each group in the response. The message field is an escaped JSON string containing the messages. For more information about the structure of the JSON string returned, see msg-prompts-formats.

If the Lambda function returns a message, Amazon Lex passes it to the client in its response.

$sel:sessionId:PostTextResponse', postTextResponse_sessionId - A unique identifier for the session.

$sel:slotToElicit:PostTextResponse', postTextResponse_slotToElicit - If the dialogState value is ElicitSlot, returns the name of the slot for which Amazon Lex is eliciting a value.

$sel:sessionAttributes:PostText', postTextResponse_sessionAttributes - A map of key-value pairs representing the session-specific context information.

$sel:httpStatus:PostTextResponse', postTextResponse_httpStatus - The response's http status code.

Response Lenses

postTextResponse_sentimentResponse :: Lens' PostTextResponse (Maybe SentimentResponse)

The sentiment expressed in and utterance.

When the bot is configured to send utterances to Amazon Comprehend for sentiment analysis, this field contains the result of the analysis.

postTextResponse_nluIntentConfidence :: Lens' PostTextResponse (Maybe IntentConfidence)

Provides a score that indicates how confident Amazon Lex is that the returned intent is the one that matches the user's intent. The score is between 0.0 and 1.0. For more information, see Confidence Scores.

The score is a relative score, not an absolute score. The score may change based on improvements to Amazon Lex.

postTextResponse_slots :: Lens' PostTextResponse (Maybe (HashMap Text Text))

The intent slots that Amazon Lex detected from the user input in the conversation.

Amazon Lex creates a resolution list containing likely values for a slot. The value that it returns is determined by the valueSelectionStrategy selected when the slot type was created or updated. If valueSelectionStrategy is set to ORIGINAL_VALUE, the value provided by the user is returned, if the user value is similar to the slot values. If valueSelectionStrategy is set to TOP_RESOLUTION Amazon Lex returns the first value in the resolution list or, if there is no resolution list, null. If you don't specify a valueSelectionStrategy, the default is ORIGINAL_VALUE.

postTextResponse_responseCard :: Lens' PostTextResponse (Maybe ResponseCard)

Represents the options that the user has to respond to the current prompt. Response Card can come from the bot configuration (in the Amazon Lex console, choose the settings button next to a slot) or from a code hook (Lambda function).

postTextResponse_intentName :: Lens' PostTextResponse (Maybe Text)

The current user intent that Amazon Lex is aware of.

postTextResponse_botVersion :: Lens' PostTextResponse (Maybe Text)

The version of the bot that responded to the conversation. You can use this information to help determine if one version of a bot is performing better than another version.

postTextResponse_dialogState :: Lens' PostTextResponse (Maybe DialogState)

Identifies the current state of the user interaction. Amazon Lex returns one of the following values as dialogState. The client can optionally use this information to customize the user interface.

ElicitIntent - Amazon Lex wants to elicit user intent.

For example, a user might utter an intent ("I want to order a pizza"). If Amazon Lex cannot infer the user intent from this utterance, it will return this dialogState.

ConfirmIntent - Amazon Lex is expecting a "yes" or "no" response.

For example, Amazon Lex wants user confirmation before fulfilling an intent.

Instead of a simple "yes" or "no," a user might respond with additional information. For example, "yes, but make it thick crust pizza" or "no, I want to order a drink". Amazon Lex can process such additional information (in these examples, update the crust type slot value, or change intent from OrderPizza to OrderDrink).

ElicitSlot - Amazon Lex is expecting a slot value for the current intent.

For example, suppose that in the response Amazon Lex sends this message: "What size pizza would you like?". A user might reply with the slot value (e.g., "medium"). The user might also provide additional information in the response (e.g., "medium thick crust pizza"). Amazon Lex can process such additional information appropriately.

Fulfilled - Conveys that the Lambda function configured for the intent has successfully fulfilled the intent.

ReadyForFulfillment - Conveys that the client has to fulfill the intent.

Failed - Conveys that the conversation with the user failed.

This can happen for various reasons including that the user did not provide an appropriate response to prompts from the service (you can configure how many times Amazon Lex can prompt a user for specific information), or the Lambda function failed to fulfill the intent.

postTextResponse_activeContexts :: Lens' PostTextResponse (Maybe [ActiveContext])

A list of active contexts for the session. A context can be set when an intent is fulfilled or by calling the PostContent, PostText, or PutSession operation.

You can use a context to control the intents that can follow up an intent, or to modify the operation of your application.

postTextResponse_alternativeIntents :: Lens' PostTextResponse (Maybe [PredictedIntent])

One to four alternative intents that may be applicable to the user's intent.

Each alternative includes a score that indicates how confident Amazon Lex is that the intent matches the user's intent. The intents are sorted by the confidence score.

postTextResponse_messageFormat :: Lens' PostTextResponse (Maybe MessageFormatType)

The format of the response message. One of the following values:

PlainText - The message contains plain UTF-8 text.

CustomPayload - The message is a custom format defined by the Lambda function.

SSML - The message contains text formatted for voice output.

Composite - The message contains an escaped JSON object containing one or more messages from the groups that messages were assigned to when the intent was created.

postTextResponse_message :: Lens' PostTextResponse (Maybe Text)

The message to convey to the user. The message can come from the bot's configuration or from a Lambda function.

If the intent is not configured with a Lambda function, or if the Lambda function returned Delegate as the dialogAction.type its response, Amazon Lex decides on the next course of action and selects an appropriate message from the bot's configuration based on the current interaction context. For example, if Amazon Lex isn't able to understand user input, it uses a clarification prompt message.

When you create an intent you can assign messages to groups. When messages are assigned to groups Amazon Lex returns one message from each group in the response. The message field is an escaped JSON string containing the messages. For more information about the structure of the JSON string returned, see msg-prompts-formats.

If the Lambda function returns a message, Amazon Lex passes it to the client in its response.

postTextResponse_sessionId :: Lens' PostTextResponse (Maybe Text)

A unique identifier for the session.

postTextResponse_slotToElicit :: Lens' PostTextResponse (Maybe Text)

If the dialogState value is ElicitSlot, returns the name of the slot for which Amazon Lex is eliciting a value.

postTextResponse_sessionAttributes :: Lens' PostTextResponse (Maybe (HashMap Text Text))

A map of key-value pairs representing the session-specific context information.

postTextResponse_httpStatus :: Lens' PostTextResponse Int

The response's http status code.