Network.AWS.LexRuntime.PutSession

Description

Creates a new session or modifies an existing session with an Amazon Lex bot. Use this operation to enable your application to set the state of the bot.

For more information, see Managing Sessions.

Creating a Request

data PutSession

See: newPutSession smart constructor.

Constructors

PutSession'

Fields

sessionAttributes :: Maybe (Sensitive (HashMap Text Text))

Map of key/value pairs representing the session-specific context information. It contains application information passed between Amazon Lex and a client application.

dialogAction :: Maybe DialogAction

Sets the next action that the bot should take to fulfill the conversation.

accept :: Maybe Text

The message that Amazon Lex returns in the response can be either text or speech based depending on the value of this field.

If the value is text/plain; charset=utf-8, Amazon Lex returns text in the response.

If the value begins with audio/, Amazon Lex returns speech in the response. Amazon Lex uses Amazon Polly to generate the speech in the configuration that you specify. For example, if you specify audio/mpeg as the value, Amazon Lex returns speech in the MPEG format.

If the value is audio/pcm, the speech is returned as audio/pcm in 16-bit, little endian format.

The following are the accepted values:

audio/mpeg

audio/ogg

audio/pcm

audio/* (defaults to mpeg)

text/plain; charset=utf-8

recentIntentSummaryView :: Maybe [IntentSummary]

A summary of the recent intents for the bot. You can use the intent summary view to set a checkpoint label on an intent and modify attributes of intents. You can also use it to remove or add intent summary objects to the list.

An intent that you modify or add to the list must make sense for the bot. For example, the intent name must be valid for the bot. You must provide valid values for:

intentName

slot names

slotToElict

If you send the recentIntentSummaryView parameter in a PutSession request, the contents of the new summary view replaces the old summary view. For example, if a GetSession request returns three intents in the summary view and you call PutSession with one intent in the summary view, the next call to GetSession will only return one intent.

activeContexts :: Maybe (Sensitive [ActiveContext])

A list of contexts active for the request. A context can be activated when a previous intent is fulfilled, or by including the context in the request,

If you don't specify a list of contexts, Amazon Lex will use the current list of contexts for the session. If you specify an empty list, all contexts for the session are cleared.

botName :: Text

The name of the bot that contains the session data.

botAlias :: Text

The alias in use for the bot that contains the session data.

userId :: Text

The ID of the client application user. Amazon Lex uses this to identify a user's conversation with your bot.

newPutSession

Arguments

:: Text
$sel:botName:PutSession'

-> Text
$sel:botAlias:PutSession'

-> Text
$sel:userId:PutSession'

-> PutSession

Create a value of PutSession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sessionAttributes:PutSession', putSession_sessionAttributes - Map of key/value pairs representing the session-specific context information. It contains application information passed between Amazon Lex and a client application.

$sel:dialogAction:PutSession', putSession_dialogAction - Sets the next action that the bot should take to fulfill the conversation.

$sel:accept:PutSession', putSession_accept - The message that Amazon Lex returns in the response can be either text or speech based depending on the value of this field.

If the value is text/plain; charset=utf-8, Amazon Lex returns text in the response.

If the value begins with audio/, Amazon Lex returns speech in the response. Amazon Lex uses Amazon Polly to generate the speech in the configuration that you specify. For example, if you specify audio/mpeg as the value, Amazon Lex returns speech in the MPEG format.

If the value is audio/pcm, the speech is returned as audio/pcm in 16-bit, little endian format.

The following are the accepted values:

audio/mpeg

audio/ogg

audio/pcm

audio/* (defaults to mpeg)

text/plain; charset=utf-8

$sel:recentIntentSummaryView:PutSession', putSession_recentIntentSummaryView - A summary of the recent intents for the bot. You can use the intent summary view to set a checkpoint label on an intent and modify attributes of intents. You can also use it to remove or add intent summary objects to the list.

An intent that you modify or add to the list must make sense for the bot. For example, the intent name must be valid for the bot. You must provide valid values for:

intentName

slot names

slotToElict

If you send the recentIntentSummaryView parameter in a PutSession request, the contents of the new summary view replaces the old summary view. For example, if a GetSession request returns three intents in the summary view and you call PutSession with one intent in the summary view, the next call to GetSession will only return one intent.

$sel:activeContexts:PutSession', putSession_activeContexts - A list of contexts active for the request. A context can be activated when a previous intent is fulfilled, or by including the context in the request,

If you don't specify a list of contexts, Amazon Lex will use the current list of contexts for the session. If you specify an empty list, all contexts for the session are cleared.

$sel:botName:PutSession', putSession_botName - The name of the bot that contains the session data.

$sel:botAlias:PutSession', putSession_botAlias - The alias in use for the bot that contains the session data.

$sel:userId:PutSession', putSession_userId - The ID of the client application user. Amazon Lex uses this to identify a user's conversation with your bot.

Request Lenses

putSession_sessionAttributes :: Lens' PutSession (Maybe (HashMap Text Text))

Map of key/value pairs representing the session-specific context information. It contains application information passed between Amazon Lex and a client application.

putSession_dialogAction :: Lens' PutSession (Maybe DialogAction)

Sets the next action that the bot should take to fulfill the conversation.

putSession_accept :: Lens' PutSession (Maybe Text)

The message that Amazon Lex returns in the response can be either text or speech based depending on the value of this field.

If the value is text/plain; charset=utf-8, Amazon Lex returns text in the response.

If the value begins with audio/, Amazon Lex returns speech in the response. Amazon Lex uses Amazon Polly to generate the speech in the configuration that you specify. For example, if you specify audio/mpeg as the value, Amazon Lex returns speech in the MPEG format.

If the value is audio/pcm, the speech is returned as audio/pcm in 16-bit, little endian format.

The following are the accepted values:

audio/mpeg

audio/ogg

audio/pcm

audio/* (defaults to mpeg)

text/plain; charset=utf-8

putSession_recentIntentSummaryView :: Lens' PutSession (Maybe [IntentSummary])

A summary of the recent intents for the bot. You can use the intent summary view to set a checkpoint label on an intent and modify attributes of intents. You can also use it to remove or add intent summary objects to the list.

An intent that you modify or add to the list must make sense for the bot. For example, the intent name must be valid for the bot. You must provide valid values for:

intentName

slot names

slotToElict

If you send the recentIntentSummaryView parameter in a PutSession request, the contents of the new summary view replaces the old summary view. For example, if a GetSession request returns three intents in the summary view and you call PutSession with one intent in the summary view, the next call to GetSession will only return one intent.

putSession_activeContexts :: Lens' PutSession (Maybe [ActiveContext])

A list of contexts active for the request. A context can be activated when a previous intent is fulfilled, or by including the context in the request,

If you don't specify a list of contexts, Amazon Lex will use the current list of contexts for the session. If you specify an empty list, all contexts for the session are cleared.

putSession_botName :: Lens' PutSession Text

The name of the bot that contains the session data.

putSession_botAlias :: Lens' PutSession Text

The alias in use for the bot that contains the session data.

putSession_userId :: Lens' PutSession Text

The ID of the client application user. Amazon Lex uses this to identify a user's conversation with your bot.

Destructuring the Response

data PutSessionResponse

See: newPutSessionResponse smart constructor.

Constructors

PutSessionResponse'

Fields

dialogState :: Maybe DialogState

ConfirmIntent - Amazon Lex is expecting a "yes" or "no" response to confirm the intent before fulfilling an intent.

ElicitIntent - Amazon Lex wants to elicit the user's intent.

ElicitSlot - Amazon Lex is expecting the value of a slot for the current intent.

Failed - Conveys that the conversation with the user has failed. This can happen for various reasons, including the user does not provide an appropriate response to prompts from the service, or if the Lambda function fails to fulfill the intent.

Fulfilled - Conveys that the Lambda function has sucessfully fulfilled the intent.

ReadyForFulfillment - Conveys that the client has to fulfill the intent.

sessionAttributes :: Maybe Text

Map of key/value pairs representing session-specific context information.

encodedMessage :: Maybe (Sensitive Text)

The next message that should be presented to the user.

The encodedMessage field is base-64 encoded. You must decode the field before you can use the value.

contentType :: Maybe Text

Content type as specified in the Accept HTTP header in the request.

intentName :: Maybe Text

The name of the current intent.

message :: Maybe (Sensitive Text)

The next message that should be presented to the user.

You can only use this field in the de-DE, en-AU, en-GB, en-US, es-419, es-ES, es-US, fr-CA, fr-FR, and it-IT locales. In all other locales, the message field is null. You should use the encodedMessage field instead.

sessionId :: Maybe Text

A unique identifier for the session.

messageFormat :: Maybe MessageFormatType

The format of the response message. One of the following values:

PlainText - The message contains plain UTF-8 text.

CustomPayload - The message is a custom format for the client.

SSML - The message contains text formatted for voice output.

Composite - The message contains an escaped JSON object containing one or more messages from the groups that messages were assigned to when the intent was created.

slots :: Maybe Text

Map of zero or more intent slots Amazon Lex detected from the user input during the conversation.

Amazon Lex creates a resolution list containing likely values for a slot. The value that it returns is determined by the valueSelectionStrategy selected when the slot type was created or updated. If valueSelectionStrategy is set to ORIGINAL_VALUE, the value provided by the user is returned, if the user value is similar to the slot values. If valueSelectionStrategy is set to TOP_RESOLUTION Amazon Lex returns the first value in the resolution list or, if there is no resolution list, null. If you don't specify a valueSelectionStrategy the default is ORIGINAL_VALUE.

slotToElicit :: Maybe Text

If the dialogState is ElicitSlot, returns the name of the slot for which Amazon Lex is eliciting a value.

activeContexts :: Maybe (Sensitive Text)

A list of active contexts for the session.

httpStatus :: Int

The response's http status code.

audioStream :: ResponseBody

The audio version of the message to convey to the user.

newPutSessionResponse

Arguments

:: Int
$sel:httpStatus:PutSessionResponse'

-> ResponseBody
$sel:audioStream:PutSessionResponse'

-> PutSessionResponse

Create a value of PutSessionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dialogState:PutSessionResponse', putSessionResponse_dialogState - - ConfirmIntent - Amazon Lex is expecting a "yes" or "no" response to confirm the intent before fulfilling an intent.

ElicitIntent - Amazon Lex wants to elicit the user's intent.

ElicitSlot - Amazon Lex is expecting the value of a slot for the current intent.

Failed - Conveys that the conversation with the user has failed. This can happen for various reasons, including the user does not provide an appropriate response to prompts from the service, or if the Lambda function fails to fulfill the intent.

Fulfilled - Conveys that the Lambda function has sucessfully fulfilled the intent.

ReadyForFulfillment - Conveys that the client has to fulfill the intent.

$sel:sessionAttributes:PutSession', putSessionResponse_sessionAttributes - Map of key/value pairs representing session-specific context information.

$sel:encodedMessage:PutSessionResponse', putSessionResponse_encodedMessage - The next message that should be presented to the user.

The encodedMessage field is base-64 encoded. You must decode the field before you can use the value.

$sel:contentType:PutSessionResponse', putSessionResponse_contentType - Content type as specified in the Accept HTTP header in the request.

$sel:intentName:PutSessionResponse', putSessionResponse_intentName - The name of the current intent.

$sel:message:PutSessionResponse', putSessionResponse_message - The next message that should be presented to the user.

You can only use this field in the de-DE, en-AU, en-GB, en-US, es-419, es-ES, es-US, fr-CA, fr-FR, and it-IT locales. In all other locales, the message field is null. You should use the encodedMessage field instead.

$sel:sessionId:PutSessionResponse', putSessionResponse_sessionId - A unique identifier for the session.

$sel:messageFormat:PutSessionResponse', putSessionResponse_messageFormat - The format of the response message. One of the following values:

PlainText - The message contains plain UTF-8 text.

CustomPayload - The message is a custom format for the client.

SSML - The message contains text formatted for voice output.

Composite - The message contains an escaped JSON object containing one or more messages from the groups that messages were assigned to when the intent was created.

$sel:slots:PutSessionResponse', putSessionResponse_slots - Map of zero or more intent slots Amazon Lex detected from the user input during the conversation.

Amazon Lex creates a resolution list containing likely values for a slot. The value that it returns is determined by the valueSelectionStrategy selected when the slot type was created or updated. If valueSelectionStrategy is set to ORIGINAL_VALUE, the value provided by the user is returned, if the user value is similar to the slot values. If valueSelectionStrategy is set to TOP_RESOLUTION Amazon Lex returns the first value in the resolution list or, if there is no resolution list, null. If you don't specify a valueSelectionStrategy the default is ORIGINAL_VALUE.

$sel:slotToElicit:PutSessionResponse', putSessionResponse_slotToElicit - If the dialogState is ElicitSlot, returns the name of the slot for which Amazon Lex is eliciting a value.

$sel:activeContexts:PutSession', putSessionResponse_activeContexts - A list of active contexts for the session.

$sel:httpStatus:PutSessionResponse', putSessionResponse_httpStatus - The response's http status code.

$sel:audioStream:PutSessionResponse', putSessionResponse_audioStream - The audio version of the message to convey to the user.

Response Lenses

putSessionResponse_dialogState :: Lens' PutSessionResponse (Maybe DialogState)

ConfirmIntent - Amazon Lex is expecting a "yes" or "no" response to confirm the intent before fulfilling an intent.

ElicitIntent - Amazon Lex wants to elicit the user's intent.

ElicitSlot - Amazon Lex is expecting the value of a slot for the current intent.

Failed - Conveys that the conversation with the user has failed. This can happen for various reasons, including the user does not provide an appropriate response to prompts from the service, or if the Lambda function fails to fulfill the intent.

Fulfilled - Conveys that the Lambda function has sucessfully fulfilled the intent.

ReadyForFulfillment - Conveys that the client has to fulfill the intent.

putSessionResponse_sessionAttributes :: Lens' PutSessionResponse (Maybe Text)

Map of key/value pairs representing session-specific context information.

putSessionResponse_encodedMessage :: Lens' PutSessionResponse (Maybe Text)

The next message that should be presented to the user.

The encodedMessage field is base-64 encoded. You must decode the field before you can use the value.

putSessionResponse_contentType :: Lens' PutSessionResponse (Maybe Text)

Content type as specified in the Accept HTTP header in the request.

putSessionResponse_intentName :: Lens' PutSessionResponse (Maybe Text)

The name of the current intent.

putSessionResponse_message :: Lens' PutSessionResponse (Maybe Text)

The next message that should be presented to the user.

You can only use this field in the de-DE, en-AU, en-GB, en-US, es-419, es-ES, es-US, fr-CA, fr-FR, and it-IT locales. In all other locales, the message field is null. You should use the encodedMessage field instead.

putSessionResponse_sessionId :: Lens' PutSessionResponse (Maybe Text)

A unique identifier for the session.

putSessionResponse_messageFormat :: Lens' PutSessionResponse (Maybe MessageFormatType)

The format of the response message. One of the following values:

PlainText - The message contains plain UTF-8 text.

CustomPayload - The message is a custom format for the client.

SSML - The message contains text formatted for voice output.

Composite - The message contains an escaped JSON object containing one or more messages from the groups that messages were assigned to when the intent was created.

putSessionResponse_slots :: Lens' PutSessionResponse (Maybe Text)

Map of zero or more intent slots Amazon Lex detected from the user input during the conversation.

Amazon Lex creates a resolution list containing likely values for a slot. The value that it returns is determined by the valueSelectionStrategy selected when the slot type was created or updated. If valueSelectionStrategy is set to ORIGINAL_VALUE, the value provided by the user is returned, if the user value is similar to the slot values. If valueSelectionStrategy is set to TOP_RESOLUTION Amazon Lex returns the first value in the resolution list or, if there is no resolution list, null. If you don't specify a valueSelectionStrategy the default is ORIGINAL_VALUE.

putSessionResponse_slotToElicit :: Lens' PutSessionResponse (Maybe Text)

If the dialogState is ElicitSlot, returns the name of the slot for which Amazon Lex is eliciting a value.

putSessionResponse_activeContexts :: Lens' PutSessionResponse (Maybe Text)

A list of active contexts for the session.

putSessionResponse_httpStatus :: Lens' PutSessionResponse Int

The response's http status code.

putSessionResponse_audioStream :: Lens' PutSessionResponse ResponseBody

The audio version of the message to convey to the user.