Network.AWS.LexRuntime.Types.GenericAttachment

Description

Documentation

data GenericAttachment

Represents an option rendered to the user when a prompt is shown. It could be an image, a button, a link, or text.

See: newGenericAttachment smart constructor.

Constructors

GenericAttachment'

Fields

title :: Maybe Text

The title of the option.

buttons :: Maybe [Button]

The list of options to show to the user.

attachmentLinkUrl :: Maybe Text

The URL of an attachment to the response card.

imageUrl :: Maybe Text

The URL of an image that is displayed to the user.

subTitle :: Maybe Text

The subtitle shown below the title.

newGenericAttachment :: GenericAttachment

Create a value of GenericAttachment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:title:GenericAttachment', genericAttachment_title - The title of the option.

$sel:buttons:GenericAttachment', genericAttachment_buttons - The list of options to show to the user.

$sel:attachmentLinkUrl:GenericAttachment', genericAttachment_attachmentLinkUrl - The URL of an attachment to the response card.

$sel:imageUrl:GenericAttachment', genericAttachment_imageUrl - The URL of an image that is displayed to the user.

$sel:subTitle:GenericAttachment', genericAttachment_subTitle - The subtitle shown below the title.

genericAttachment_title :: Lens' GenericAttachment (Maybe Text)

The title of the option.

genericAttachment_buttons :: Lens' GenericAttachment (Maybe [Button])

The list of options to show to the user.

genericAttachment_attachmentLinkUrl :: Lens' GenericAttachment (Maybe Text)

The URL of an attachment to the response card.

genericAttachment_imageUrl :: Lens' GenericAttachment (Maybe Text)

The URL of an image that is displayed to the user.

genericAttachment_subTitle :: Lens' GenericAttachment (Maybe Text)

The subtitle shown below the title.