Network.AWS.LexRuntime.Types.IntentSummary

Description

Documentation

data IntentSummary

Provides information about the state of an intent. You can use this information to get the current state of an intent so that you can process the intent, or so that you can return the intent to its previous state.

See: newIntentSummary smart constructor.

Constructors

IntentSummary'

Fields

intentName :: Maybe Text

The name of the intent.

fulfillmentState :: Maybe FulfillmentState

The fulfillment state of the intent. The possible values are:

Failed - The Lambda function associated with the intent failed to fulfill the intent.

Fulfilled - The intent has fulfilled by the Lambda function associated with the intent.

ReadyForFulfillment - All of the information necessary for the intent is present and the intent ready to be fulfilled by the client application.

slots :: Maybe (Sensitive (HashMap Text Text))

Map of the slots that have been gathered and their values.

checkpointLabel :: Maybe Text

A user-defined label that identifies a particular intent. You can use this label to return to a previous intent.

Use the checkpointLabelFilter parameter of the GetSessionRequest operation to filter the intents returned by the operation to those with only the specified label.

slotToElicit :: Maybe Text

The next slot to elicit from the user. If there is not slot to elicit, the field is blank.

confirmationStatus :: Maybe ConfirmationStatus

The status of the intent after the user responds to the confirmation prompt. If the user confirms the intent, Amazon Lex sets this field to Confirmed. If the user denies the intent, Amazon Lex sets this value to Denied. The possible values are:

Confirmed - The user has responded "Yes" to the confirmation prompt, confirming that the intent is complete and that it is ready to be fulfilled.

Denied - The user has responded "No" to the confirmation prompt.

None - The user has never been prompted for confirmation; or, the user was prompted but did not confirm or deny the prompt.

dialogActionType :: DialogActionType

The next action that the bot should take in its interaction with the user. The possible values are:

ConfirmIntent - The next action is asking the user if the intent is complete and ready to be fulfilled. This is a yes/no question such as "Place the order?"

Close - Indicates that the there will not be a response from the user. For example, the statement "Your order has been placed" does not require a response.

ElicitIntent - The next action is to determine the intent that the user wants to fulfill.

ElicitSlot - The next action is to elicit a slot value from the user.

newIntentSummary

Arguments

:: DialogActionType
$sel:dialogActionType:IntentSummary'

-> IntentSummary

Create a value of IntentSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:intentName:IntentSummary', intentSummary_intentName - The name of the intent.

$sel:fulfillmentState:IntentSummary', intentSummary_fulfillmentState - The fulfillment state of the intent. The possible values are:

Failed - The Lambda function associated with the intent failed to fulfill the intent.

Fulfilled - The intent has fulfilled by the Lambda function associated with the intent.

ReadyForFulfillment - All of the information necessary for the intent is present and the intent ready to be fulfilled by the client application.

$sel:slots:IntentSummary', intentSummary_slots - Map of the slots that have been gathered and their values.

$sel:checkpointLabel:IntentSummary', intentSummary_checkpointLabel - A user-defined label that identifies a particular intent. You can use this label to return to a previous intent.

Use the checkpointLabelFilter parameter of the GetSessionRequest operation to filter the intents returned by the operation to those with only the specified label.

$sel:slotToElicit:IntentSummary', intentSummary_slotToElicit - The next slot to elicit from the user. If there is not slot to elicit, the field is blank.

$sel:confirmationStatus:IntentSummary', intentSummary_confirmationStatus - The status of the intent after the user responds to the confirmation prompt. If the user confirms the intent, Amazon Lex sets this field to Confirmed. If the user denies the intent, Amazon Lex sets this value to Denied. The possible values are:

Confirmed - The user has responded "Yes" to the confirmation prompt, confirming that the intent is complete and that it is ready to be fulfilled.

Denied - The user has responded "No" to the confirmation prompt.

None - The user has never been prompted for confirmation; or, the user was prompted but did not confirm or deny the prompt.

$sel:dialogActionType:IntentSummary', intentSummary_dialogActionType - The next action that the bot should take in its interaction with the user. The possible values are:

ConfirmIntent - The next action is asking the user if the intent is complete and ready to be fulfilled. This is a yes/no question such as "Place the order?"

Close - Indicates that the there will not be a response from the user. For example, the statement "Your order has been placed" does not require a response.

ElicitIntent - The next action is to determine the intent that the user wants to fulfill.

ElicitSlot - The next action is to elicit a slot value from the user.

intentSummary_intentName :: Lens' IntentSummary (Maybe Text)

The name of the intent.

intentSummary_fulfillmentState :: Lens' IntentSummary (Maybe FulfillmentState)

The fulfillment state of the intent. The possible values are:

Failed - The Lambda function associated with the intent failed to fulfill the intent.

Fulfilled - The intent has fulfilled by the Lambda function associated with the intent.

ReadyForFulfillment - All of the information necessary for the intent is present and the intent ready to be fulfilled by the client application.

intentSummary_slots :: Lens' IntentSummary (Maybe (HashMap Text Text))

Map of the slots that have been gathered and their values.

intentSummary_checkpointLabel :: Lens' IntentSummary (Maybe Text)

A user-defined label that identifies a particular intent. You can use this label to return to a previous intent.

Use the checkpointLabelFilter parameter of the GetSessionRequest operation to filter the intents returned by the operation to those with only the specified label.

intentSummary_slotToElicit :: Lens' IntentSummary (Maybe Text)

The next slot to elicit from the user. If there is not slot to elicit, the field is blank.

intentSummary_confirmationStatus :: Lens' IntentSummary (Maybe ConfirmationStatus)

The status of the intent after the user responds to the confirmation prompt. If the user confirms the intent, Amazon Lex sets this field to Confirmed. If the user denies the intent, Amazon Lex sets this value to Denied. The possible values are:

Confirmed - The user has responded "Yes" to the confirmation prompt, confirming that the intent is complete and that it is ready to be fulfilled.

Denied - The user has responded "No" to the confirmation prompt.

None - The user has never been prompted for confirmation; or, the user was prompted but did not confirm or deny the prompt.

intentSummary_dialogActionType :: Lens' IntentSummary DialogActionType

The next action that the bot should take in its interaction with the user. The possible values are:

ConfirmIntent - The next action is asking the user if the intent is complete and ready to be fulfilled. This is a yes/no question such as "Place the order?"

Close - Indicates that the there will not be a response from the user. For example, the statement "Your order has been placed" does not require a response.

ElicitIntent - The next action is to determine the intent that the user wants to fulfill.

ElicitSlot - The next action is to elicit a slot value from the user.