Network.AWS.LexRuntime.Types.PredictedIntent

Description

Documentation

data PredictedIntent

An intent that Amazon Lex suggests satisfies the user's intent. Includes the name of the intent, the confidence that Amazon Lex has that the user's intent is satisfied, and the slots defined for the intent.

See: newPredictedIntent smart constructor.

Constructors

PredictedIntent'

Fields

nluIntentConfidence :: Maybe IntentConfidence

Indicates how confident Amazon Lex is that an intent satisfies the user's intent.

slots :: Maybe (Sensitive (HashMap Text Text))

The slot and slot values associated with the predicted intent.

intentName :: Maybe Text

The name of the intent that Amazon Lex suggests satisfies the user's intent.

newPredictedIntent :: PredictedIntent

Create a value of PredictedIntent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nluIntentConfidence:PredictedIntent', predictedIntent_nluIntentConfidence - Indicates how confident Amazon Lex is that an intent satisfies the user's intent.

$sel:slots:PredictedIntent', predictedIntent_slots - The slot and slot values associated with the predicted intent.

$sel:intentName:PredictedIntent', predictedIntent_intentName - The name of the intent that Amazon Lex suggests satisfies the user's intent.

predictedIntent_nluIntentConfidence :: Lens' PredictedIntent (Maybe IntentConfidence)

Indicates how confident Amazon Lex is that an intent satisfies the user's intent.

predictedIntent_slots :: Lens' PredictedIntent (Maybe (HashMap Text Text))

The slot and slot values associated with the predicted intent.

predictedIntent_intentName :: Lens' PredictedIntent (Maybe Text)

The name of the intent that Amazon Lex suggests satisfies the user's intent.