Network.AWS.LicenseManager.ExtendLicenseConsumption

Description

Extends the expiration date for license consumption.

Creating a Request

data ExtendLicenseConsumption

See: newExtendLicenseConsumption smart constructor.

Constructors

ExtendLicenseConsumption'

Fields

dryRun :: Maybe Bool

Checks whether you have the required permissions for the action, without actually making the request. Provides an error response if you do not have the required permissions.

licenseConsumptionToken :: Text

License consumption token.

newExtendLicenseConsumption

Arguments

:: Text
$sel:licenseConsumptionToken:ExtendLicenseConsumption'

-> ExtendLicenseConsumption

Create a value of ExtendLicenseConsumption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:ExtendLicenseConsumption', extendLicenseConsumption_dryRun - Checks whether you have the required permissions for the action, without actually making the request. Provides an error response if you do not have the required permissions.

$sel:licenseConsumptionToken:ExtendLicenseConsumption', extendLicenseConsumption_licenseConsumptionToken - License consumption token.

Request Lenses

extendLicenseConsumption_dryRun :: Lens' ExtendLicenseConsumption (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request. Provides an error response if you do not have the required permissions.

extendLicenseConsumption_licenseConsumptionToken :: Lens' ExtendLicenseConsumption Text

License consumption token.

Destructuring the Response

data ExtendLicenseConsumptionResponse

See: newExtendLicenseConsumptionResponse smart constructor.

Constructors

ExtendLicenseConsumptionResponse'

Fields

expiration :: Maybe Text

Date and time at which the license consumption expires.

licenseConsumptionToken :: Maybe Text

License consumption token.

httpStatus :: Int

The response's http status code.

newExtendLicenseConsumptionResponse

Arguments

:: Int
$sel:httpStatus:ExtendLicenseConsumptionResponse'

-> ExtendLicenseConsumptionResponse

Create a value of ExtendLicenseConsumptionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:expiration:ExtendLicenseConsumptionResponse', extendLicenseConsumptionResponse_expiration - Date and time at which the license consumption expires.

$sel:licenseConsumptionToken:ExtendLicenseConsumption', extendLicenseConsumptionResponse_licenseConsumptionToken - License consumption token.

$sel:httpStatus:ExtendLicenseConsumptionResponse', extendLicenseConsumptionResponse_httpStatus - The response's http status code.

Response Lenses

extendLicenseConsumptionResponse_expiration :: Lens' ExtendLicenseConsumptionResponse (Maybe Text)

Date and time at which the license consumption expires.

extendLicenseConsumptionResponse_licenseConsumptionToken :: Lens' ExtendLicenseConsumptionResponse (Maybe Text)

License consumption token.

extendLicenseConsumptionResponse_httpStatus :: Lens' ExtendLicenseConsumptionResponse Int

The response's http status code.