Network.AWS.LicenseManager.GetLicenseConfiguration

Description

Gets detailed information about the specified license configuration.

Creating a Request

data GetLicenseConfiguration

See: newGetLicenseConfiguration smart constructor.

Constructors

GetLicenseConfiguration'

Fields

licenseConfigurationArn :: Text

Amazon Resource Name (ARN) of the license configuration.

newGetLicenseConfiguration

Arguments

:: Text
$sel:licenseConfigurationArn:GetLicenseConfiguration'

-> GetLicenseConfiguration

Create a value of GetLicenseConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:licenseConfigurationArn:GetLicenseConfiguration', getLicenseConfiguration_licenseConfigurationArn - Amazon Resource Name (ARN) of the license configuration.

Request Lenses

getLicenseConfiguration_licenseConfigurationArn :: Lens' GetLicenseConfiguration Text

Amazon Resource Name (ARN) of the license configuration.

Destructuring the Response

data GetLicenseConfigurationResponse

See: newGetLicenseConfigurationResponse smart constructor.

Constructors

GetLicenseConfigurationResponse'

Fields

status :: Maybe Text

License configuration status.

ownerAccountId :: Maybe Text

Account ID of the owner of the license configuration.

consumedLicenseSummaryList :: Maybe [ConsumedLicenseSummary]

Summaries of the licenses consumed by resources.

licenseCount :: Maybe Integer

Number of available licenses.

managedResourceSummaryList :: Maybe [ManagedResourceSummary]

Summaries of the managed resources.

name :: Maybe Text

Name of the license configuration.

licenseCountHardLimit :: Maybe Bool

Sets the number of available licenses as a hard limit.

disassociateWhenNotFound :: Maybe Bool

When true, disassociates a resource when software is uninstalled.

productInformationList :: Maybe [ProductInformation]

Product information.

licenseCountingType :: Maybe LicenseCountingType

Dimension for which the licenses are counted.

automatedDiscoveryInformation :: Maybe AutomatedDiscoveryInformation

Automated discovery information.

consumedLicenses :: Maybe Integer

Number of licenses assigned to resources.

licenseRules :: Maybe [Text]

License rules.

licenseConfigurationId :: Maybe Text

Unique ID for the license configuration.

description :: Maybe Text

Description of the license configuration.

licenseConfigurationArn :: Maybe Text

Amazon Resource Name (ARN) of the license configuration.

tags :: Maybe [Tag]

Tags for the license configuration.

httpStatus :: Int

The response's http status code.

newGetLicenseConfigurationResponse

Arguments

:: Int
$sel:httpStatus:GetLicenseConfigurationResponse'

-> GetLicenseConfigurationResponse

Create a value of GetLicenseConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetLicenseConfigurationResponse', getLicenseConfigurationResponse_status - License configuration status.

$sel:ownerAccountId:GetLicenseConfigurationResponse', getLicenseConfigurationResponse_ownerAccountId - Account ID of the owner of the license configuration.

$sel:consumedLicenseSummaryList:GetLicenseConfigurationResponse', getLicenseConfigurationResponse_consumedLicenseSummaryList - Summaries of the licenses consumed by resources.

$sel:licenseCount:GetLicenseConfigurationResponse', getLicenseConfigurationResponse_licenseCount - Number of available licenses.

$sel:managedResourceSummaryList:GetLicenseConfigurationResponse', getLicenseConfigurationResponse_managedResourceSummaryList - Summaries of the managed resources.

$sel:name:GetLicenseConfigurationResponse', getLicenseConfigurationResponse_name - Name of the license configuration.

$sel:licenseCountHardLimit:GetLicenseConfigurationResponse', getLicenseConfigurationResponse_licenseCountHardLimit - Sets the number of available licenses as a hard limit.

$sel:disassociateWhenNotFound:GetLicenseConfigurationResponse', getLicenseConfigurationResponse_disassociateWhenNotFound - When true, disassociates a resource when software is uninstalled.

$sel:productInformationList:GetLicenseConfigurationResponse', getLicenseConfigurationResponse_productInformationList - Product information.

$sel:licenseCountingType:GetLicenseConfigurationResponse', getLicenseConfigurationResponse_licenseCountingType - Dimension for which the licenses are counted.

$sel:automatedDiscoveryInformation:GetLicenseConfigurationResponse', getLicenseConfigurationResponse_automatedDiscoveryInformation - Automated discovery information.

$sel:consumedLicenses:GetLicenseConfigurationResponse', getLicenseConfigurationResponse_consumedLicenses - Number of licenses assigned to resources.

$sel:licenseRules:GetLicenseConfigurationResponse', getLicenseConfigurationResponse_licenseRules - License rules.

$sel:licenseConfigurationId:GetLicenseConfigurationResponse', getLicenseConfigurationResponse_licenseConfigurationId - Unique ID for the license configuration.

$sel:description:GetLicenseConfigurationResponse', getLicenseConfigurationResponse_description - Description of the license configuration.

$sel:licenseConfigurationArn:GetLicenseConfiguration', getLicenseConfigurationResponse_licenseConfigurationArn - Amazon Resource Name (ARN) of the license configuration.

$sel:tags:GetLicenseConfigurationResponse', getLicenseConfigurationResponse_tags - Tags for the license configuration.

$sel:httpStatus:GetLicenseConfigurationResponse', getLicenseConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

getLicenseConfigurationResponse_status :: Lens' GetLicenseConfigurationResponse (Maybe Text)

License configuration status.

getLicenseConfigurationResponse_ownerAccountId :: Lens' GetLicenseConfigurationResponse (Maybe Text)

Account ID of the owner of the license configuration.

getLicenseConfigurationResponse_consumedLicenseSummaryList :: Lens' GetLicenseConfigurationResponse (Maybe [ConsumedLicenseSummary])

Summaries of the licenses consumed by resources.

getLicenseConfigurationResponse_licenseCount :: Lens' GetLicenseConfigurationResponse (Maybe Integer)

Number of available licenses.

getLicenseConfigurationResponse_managedResourceSummaryList :: Lens' GetLicenseConfigurationResponse (Maybe [ManagedResourceSummary])

Summaries of the managed resources.

getLicenseConfigurationResponse_name :: Lens' GetLicenseConfigurationResponse (Maybe Text)

Name of the license configuration.

getLicenseConfigurationResponse_licenseCountHardLimit :: Lens' GetLicenseConfigurationResponse (Maybe Bool)

Sets the number of available licenses as a hard limit.

getLicenseConfigurationResponse_disassociateWhenNotFound :: Lens' GetLicenseConfigurationResponse (Maybe Bool)

When true, disassociates a resource when software is uninstalled.

getLicenseConfigurationResponse_productInformationList :: Lens' GetLicenseConfigurationResponse (Maybe [ProductInformation])

Product information.

getLicenseConfigurationResponse_licenseCountingType :: Lens' GetLicenseConfigurationResponse (Maybe LicenseCountingType)

Dimension for which the licenses are counted.

getLicenseConfigurationResponse_automatedDiscoveryInformation :: Lens' GetLicenseConfigurationResponse (Maybe AutomatedDiscoveryInformation)

Automated discovery information.

getLicenseConfigurationResponse_consumedLicenses :: Lens' GetLicenseConfigurationResponse (Maybe Integer)

Number of licenses assigned to resources.

getLicenseConfigurationResponse_licenseRules :: Lens' GetLicenseConfigurationResponse (Maybe [Text])

License rules.

getLicenseConfigurationResponse_licenseConfigurationId :: Lens' GetLicenseConfigurationResponse (Maybe Text)

Unique ID for the license configuration.

getLicenseConfigurationResponse_description :: Lens' GetLicenseConfigurationResponse (Maybe Text)

Description of the license configuration.

getLicenseConfigurationResponse_licenseConfigurationArn :: Lens' GetLicenseConfigurationResponse (Maybe Text)

Amazon Resource Name (ARN) of the license configuration.

getLicenseConfigurationResponse_tags :: Lens' GetLicenseConfigurationResponse (Maybe [Tag])

Tags for the license configuration.

getLicenseConfigurationResponse_httpStatus :: Lens' GetLicenseConfigurationResponse Int

The response's http status code.