Network.AWS.LicenseManager.ListFailuresForLicenseConfigurationOperations

Description

Lists the license configuration operations that failed.

Creating a Request

data ListFailuresForLicenseConfigurationOperations

See: newListFailuresForLicenseConfigurationOperations smart constructor.

Constructors

ListFailuresForLicenseConfigurationOperations'

Fields

nextToken :: Maybe Text

Token for the next set of results.

maxResults :: Maybe Int

Maximum number of results to return in a single call.

licenseConfigurationArn :: Text

Amazon Resource Name of the license configuration.

newListFailuresForLicenseConfigurationOperations

Arguments

:: Text
$sel:licenseConfigurationArn:ListFailuresForLicenseConfigurationOperations'

-> ListFailuresForLicenseConfigurationOperations

Create a value of ListFailuresForLicenseConfigurationOperations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFailuresForLicenseConfigurationOperations', listFailuresForLicenseConfigurationOperations_nextToken - Token for the next set of results.

$sel:maxResults:ListFailuresForLicenseConfigurationOperations', listFailuresForLicenseConfigurationOperations_maxResults - Maximum number of results to return in a single call.

$sel:licenseConfigurationArn:ListFailuresForLicenseConfigurationOperations', listFailuresForLicenseConfigurationOperations_licenseConfigurationArn - Amazon Resource Name of the license configuration.

Request Lenses

listFailuresForLicenseConfigurationOperations_nextToken :: Lens' ListFailuresForLicenseConfigurationOperations (Maybe Text)

Token for the next set of results.

listFailuresForLicenseConfigurationOperations_maxResults :: Lens' ListFailuresForLicenseConfigurationOperations (Maybe Int)

Maximum number of results to return in a single call.

listFailuresForLicenseConfigurationOperations_licenseConfigurationArn :: Lens' ListFailuresForLicenseConfigurationOperations Text

Amazon Resource Name of the license configuration.

Destructuring the Response

data ListFailuresForLicenseConfigurationOperationsResponse

See: newListFailuresForLicenseConfigurationOperationsResponse smart constructor.

Constructors

ListFailuresForLicenseConfigurationOperationsResponse'

Fields

nextToken :: Maybe Text

Token for the next set of results.

licenseOperationFailureList :: Maybe [LicenseOperationFailure]

License configuration operations that failed.

httpStatus :: Int

The response's http status code.

newListFailuresForLicenseConfigurationOperationsResponse

Arguments

:: Int
$sel:httpStatus:ListFailuresForLicenseConfigurationOperationsResponse'

-> ListFailuresForLicenseConfigurationOperationsResponse

Create a value of ListFailuresForLicenseConfigurationOperationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFailuresForLicenseConfigurationOperations', listFailuresForLicenseConfigurationOperationsResponse_nextToken - Token for the next set of results.

$sel:licenseOperationFailureList:ListFailuresForLicenseConfigurationOperationsResponse', listFailuresForLicenseConfigurationOperationsResponse_licenseOperationFailureList - License configuration operations that failed.

$sel:httpStatus:ListFailuresForLicenseConfigurationOperationsResponse', listFailuresForLicenseConfigurationOperationsResponse_httpStatus - The response's http status code.

Response Lenses

listFailuresForLicenseConfigurationOperationsResponse_nextToken :: Lens' ListFailuresForLicenseConfigurationOperationsResponse (Maybe Text)

Token for the next set of results.

listFailuresForLicenseConfigurationOperationsResponse_licenseOperationFailureList :: Lens' ListFailuresForLicenseConfigurationOperationsResponse (Maybe [LicenseOperationFailure])

License configuration operations that failed.

listFailuresForLicenseConfigurationOperationsResponse_httpStatus :: Lens' ListFailuresForLicenseConfigurationOperationsResponse Int

The response's http status code.