Network.AWS.LicenseManager.ListLicenseConversionTasks

Description

Lists the license type conversion tasks for your account.

Creating a Request

data ListLicenseConversionTasks

See: newListLicenseConversionTasks smart constructor.

Constructors

ListLicenseConversionTasks'

Fields

filters :: Maybe [Filter]

Filters to scope the results. Valid filters are ResourceArns and Status.

nextToken :: Maybe Text

Token for the next set of results.

maxResults :: Maybe Int

Maximum number of results to return in a single call.

newListLicenseConversionTasks :: ListLicenseConversionTasks

Create a value of ListLicenseConversionTasks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListLicenseConversionTasks', listLicenseConversionTasks_filters - Filters to scope the results. Valid filters are ResourceArns and Status.

$sel:nextToken:ListLicenseConversionTasks', listLicenseConversionTasks_nextToken - Token for the next set of results.

$sel:maxResults:ListLicenseConversionTasks', listLicenseConversionTasks_maxResults - Maximum number of results to return in a single call.

Request Lenses

listLicenseConversionTasks_filters :: Lens' ListLicenseConversionTasks (Maybe [Filter])

Filters to scope the results. Valid filters are ResourceArns and Status.

listLicenseConversionTasks_nextToken :: Lens' ListLicenseConversionTasks (Maybe Text)

Token for the next set of results.

listLicenseConversionTasks_maxResults :: Lens' ListLicenseConversionTasks (Maybe Int)

Maximum number of results to return in a single call.

Destructuring the Response

data ListLicenseConversionTasksResponse

See: newListLicenseConversionTasksResponse smart constructor.

Constructors

ListLicenseConversionTasksResponse'

Fields

licenseConversionTasks :: Maybe [LicenseConversionTask]

Information about the license configuration tasks for your account.

nextToken :: Maybe Text

Token for the next set of results.

httpStatus :: Int

The response's http status code.

newListLicenseConversionTasksResponse

Arguments

:: Int
$sel:httpStatus:ListLicenseConversionTasksResponse'

-> ListLicenseConversionTasksResponse

Create a value of ListLicenseConversionTasksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:licenseConversionTasks:ListLicenseConversionTasksResponse', listLicenseConversionTasksResponse_licenseConversionTasks - Information about the license configuration tasks for your account.

$sel:nextToken:ListLicenseConversionTasks', listLicenseConversionTasksResponse_nextToken - Token for the next set of results.

$sel:httpStatus:ListLicenseConversionTasksResponse', listLicenseConversionTasksResponse_httpStatus - The response's http status code.

Response Lenses

listLicenseConversionTasksResponse_licenseConversionTasks :: Lens' ListLicenseConversionTasksResponse (Maybe [LicenseConversionTask])

Information about the license configuration tasks for your account.

listLicenseConversionTasksResponse_nextToken :: Lens' ListLicenseConversionTasksResponse (Maybe Text)

Token for the next set of results.

listLicenseConversionTasksResponse_httpStatus :: Lens' ListLicenseConversionTasksResponse Int

The response's http status code.