Network.AWS.LicenseManager.ListLicenseVersions

Description

Lists all versions of the specified license.

Creating a Request

data ListLicenseVersions

See: newListLicenseVersions smart constructor.

Constructors

ListLicenseVersions'

Fields

nextToken :: Maybe Text

Token for the next set of results.

maxResults :: Maybe Natural

Maximum number of results to return in a single call.

licenseArn :: Text

Amazon Resource Name (ARN) of the license.

newListLicenseVersions

Arguments

:: Text
$sel:licenseArn:ListLicenseVersions'

-> ListLicenseVersions

Create a value of ListLicenseVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListLicenseVersions', listLicenseVersions_nextToken - Token for the next set of results.

$sel:maxResults:ListLicenseVersions', listLicenseVersions_maxResults - Maximum number of results to return in a single call.

$sel:licenseArn:ListLicenseVersions', listLicenseVersions_licenseArn - Amazon Resource Name (ARN) of the license.

Request Lenses

listLicenseVersions_nextToken :: Lens' ListLicenseVersions (Maybe Text)

Token for the next set of results.

listLicenseVersions_maxResults :: Lens' ListLicenseVersions (Maybe Natural)

Maximum number of results to return in a single call.

listLicenseVersions_licenseArn :: Lens' ListLicenseVersions Text

Amazon Resource Name (ARN) of the license.

Destructuring the Response

data ListLicenseVersionsResponse

See: newListLicenseVersionsResponse smart constructor.

Constructors

ListLicenseVersionsResponse'

Fields

nextToken :: Maybe Text

Token for the next set of results.

licenses :: Maybe [License]

License details.

httpStatus :: Int

The response's http status code.

newListLicenseVersionsResponse

Arguments

:: Int
$sel:httpStatus:ListLicenseVersionsResponse'

-> ListLicenseVersionsResponse

Create a value of ListLicenseVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListLicenseVersions', listLicenseVersionsResponse_nextToken - Token for the next set of results.

$sel:licenses:ListLicenseVersionsResponse', listLicenseVersionsResponse_licenses - License details.

$sel:httpStatus:ListLicenseVersionsResponse', listLicenseVersionsResponse_httpStatus - The response's http status code.

Response Lenses

listLicenseVersionsResponse_nextToken :: Lens' ListLicenseVersionsResponse (Maybe Text)

Token for the next set of results.

listLicenseVersionsResponse_licenses :: Lens' ListLicenseVersionsResponse (Maybe [License])

License details.

listLicenseVersionsResponse_httpStatus :: Lens' ListLicenseVersionsResponse Int

The response's http status code.