Network.AWS.LicenseManager.ListLicenses

Description

Lists the licenses for your account.

Creating a Request

data ListLicenses

See: newListLicenses smart constructor.

Constructors

ListLicenses'

Fields

filters :: Maybe [Filter]

Filters to scope the results. The following filters are supported:

Beneficiary

ProductSKU

Fingerprint

Status

nextToken :: Maybe Text

Token for the next set of results.

licenseArns :: Maybe [Text]

Amazon Resource Names (ARNs) of the licenses.

maxResults :: Maybe Natural

Maximum number of results to return in a single call.

newListLicenses :: ListLicenses

Create a value of ListLicenses with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListLicenses', listLicenses_filters - Filters to scope the results. The following filters are supported:

Beneficiary

ProductSKU

Fingerprint

Status

$sel:nextToken:ListLicenses', listLicenses_nextToken - Token for the next set of results.

$sel:licenseArns:ListLicenses', listLicenses_licenseArns - Amazon Resource Names (ARNs) of the licenses.

$sel:maxResults:ListLicenses', listLicenses_maxResults - Maximum number of results to return in a single call.

Request Lenses

listLicenses_filters :: Lens' ListLicenses (Maybe [Filter])

Filters to scope the results. The following filters are supported:

Beneficiary

ProductSKU

Fingerprint

Status

listLicenses_nextToken :: Lens' ListLicenses (Maybe Text)

Token for the next set of results.

listLicenses_licenseArns :: Lens' ListLicenses (Maybe [Text])

Amazon Resource Names (ARNs) of the licenses.

listLicenses_maxResults :: Lens' ListLicenses (Maybe Natural)

Maximum number of results to return in a single call.

Destructuring the Response

data ListLicensesResponse

See: newListLicensesResponse smart constructor.

Constructors

ListLicensesResponse'

Fields

nextToken :: Maybe Text

Token for the next set of results.

licenses :: Maybe [License]

License details.

httpStatus :: Int

The response's http status code.

newListLicensesResponse

Arguments

:: Int
$sel:httpStatus:ListLicensesResponse'

-> ListLicensesResponse

Create a value of ListLicensesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListLicenses', listLicensesResponse_nextToken - Token for the next set of results.

$sel:licenses:ListLicensesResponse', listLicensesResponse_licenses - License details.

$sel:httpStatus:ListLicensesResponse', listLicensesResponse_httpStatus - The response's http status code.

Response Lenses

listLicensesResponse_nextToken :: Lens' ListLicensesResponse (Maybe Text)

Token for the next set of results.

listLicensesResponse_licenses :: Lens' ListLicensesResponse (Maybe [License])

License details.

listLicensesResponse_httpStatus :: Lens' ListLicensesResponse Int

The response's http status code.