Network.AWS.LicenseManager.ListUsageForLicenseConfiguration

Description

Lists all license usage records for a license configuration, displaying license consumption details by resource at a selected point in time. Use this action to audit the current license consumption for any license inventory and configuration.

This operation returns paginated results.

Creating a Request

data ListUsageForLicenseConfiguration

See: newListUsageForLicenseConfiguration smart constructor.

Constructors

ListUsageForLicenseConfiguration'

Fields

filters :: Maybe [Filter]

Filters to scope the results. The following filters and logical operators are supported:

resourceArn - The ARN of the license configuration resource. Logical operators are EQUALS | NOT_EQUALS.

resourceType - The resource type (EC2_INSTANCE | EC2_HOST | EC2_AMI | SYSTEMS_MANAGER_MANAGED_INSTANCE). Logical operators are EQUALS | NOT_EQUALS.

resourceAccount - The ID of the account that owns the resource. Logical operators are EQUALS | NOT_EQUALS.

nextToken :: Maybe Text

Token for the next set of results.

maxResults :: Maybe Int

Maximum number of results to return in a single call.

licenseConfigurationArn :: Text

Amazon Resource Name (ARN) of the license configuration.

newListUsageForLicenseConfiguration

Arguments

:: Text
$sel:licenseConfigurationArn:ListUsageForLicenseConfiguration'

-> ListUsageForLicenseConfiguration

Create a value of ListUsageForLicenseConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:ListUsageForLicenseConfiguration', listUsageForLicenseConfiguration_filters - Filters to scope the results. The following filters and logical operators are supported:

resourceArn - The ARN of the license configuration resource. Logical operators are EQUALS | NOT_EQUALS.

resourceType - The resource type (EC2_INSTANCE | EC2_HOST | EC2_AMI | SYSTEMS_MANAGER_MANAGED_INSTANCE). Logical operators are EQUALS | NOT_EQUALS.

resourceAccount - The ID of the account that owns the resource. Logical operators are EQUALS | NOT_EQUALS.

$sel:nextToken:ListUsageForLicenseConfiguration', listUsageForLicenseConfiguration_nextToken - Token for the next set of results.

$sel:maxResults:ListUsageForLicenseConfiguration', listUsageForLicenseConfiguration_maxResults - Maximum number of results to return in a single call.

$sel:licenseConfigurationArn:ListUsageForLicenseConfiguration', listUsageForLicenseConfiguration_licenseConfigurationArn - Amazon Resource Name (ARN) of the license configuration.

Request Lenses

listUsageForLicenseConfiguration_filters :: Lens' ListUsageForLicenseConfiguration (Maybe [Filter])

Filters to scope the results. The following filters and logical operators are supported:

resourceArn - The ARN of the license configuration resource. Logical operators are EQUALS | NOT_EQUALS.

resourceType - The resource type (EC2_INSTANCE | EC2_HOST | EC2_AMI | SYSTEMS_MANAGER_MANAGED_INSTANCE). Logical operators are EQUALS | NOT_EQUALS.

resourceAccount - The ID of the account that owns the resource. Logical operators are EQUALS | NOT_EQUALS.

listUsageForLicenseConfiguration_nextToken :: Lens' ListUsageForLicenseConfiguration (Maybe Text)

Token for the next set of results.

listUsageForLicenseConfiguration_maxResults :: Lens' ListUsageForLicenseConfiguration (Maybe Int)

Maximum number of results to return in a single call.

listUsageForLicenseConfiguration_licenseConfigurationArn :: Lens' ListUsageForLicenseConfiguration Text

Amazon Resource Name (ARN) of the license configuration.

Destructuring the Response

data ListUsageForLicenseConfigurationResponse

See: newListUsageForLicenseConfigurationResponse smart constructor.

Constructors

ListUsageForLicenseConfigurationResponse'

Fields

nextToken :: Maybe Text

Token for the next set of results.

licenseConfigurationUsageList :: Maybe [LicenseConfigurationUsage]

Information about the license configurations.

httpStatus :: Int

The response's http status code.

newListUsageForLicenseConfigurationResponse

Arguments

:: Int
$sel:httpStatus:ListUsageForLicenseConfigurationResponse'

-> ListUsageForLicenseConfigurationResponse

Create a value of ListUsageForLicenseConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListUsageForLicenseConfiguration', listUsageForLicenseConfigurationResponse_nextToken - Token for the next set of results.

$sel:licenseConfigurationUsageList:ListUsageForLicenseConfigurationResponse', listUsageForLicenseConfigurationResponse_licenseConfigurationUsageList - Information about the license configurations.

$sel:httpStatus:ListUsageForLicenseConfigurationResponse', listUsageForLicenseConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

listUsageForLicenseConfigurationResponse_nextToken :: Lens' ListUsageForLicenseConfigurationResponse (Maybe Text)

Token for the next set of results.

listUsageForLicenseConfigurationResponse_licenseConfigurationUsageList :: Lens' ListUsageForLicenseConfigurationResponse (Maybe [LicenseConfigurationUsage])

Information about the license configurations.

listUsageForLicenseConfigurationResponse_httpStatus :: Lens' ListUsageForLicenseConfigurationResponse Int

The response's http status code.