Network.AWS.LicenseManager.Types.License

Description

Documentation

data License

Software license that is managed in License Manager.

See: newLicense smart constructor.

Constructors

License'

Fields

status :: Maybe LicenseStatus

License status.

productName :: Maybe Text

Product name.

licenseName :: Maybe Text

License name.

productSKU :: Maybe Text

Product SKU.

homeRegion :: Maybe Text

Home Region of the license.

version :: Maybe Text

License version.

entitlements :: Maybe [Entitlement]

License entitlements.

validity :: Maybe DatetimeRange

Date and time range during which the license is valid, in ISO8601-UTC format.

consumptionConfiguration :: Maybe ConsumptionConfiguration

Configuration for consumption of the license.

beneficiary :: Maybe Text

License beneficiary.

issuer :: Maybe IssuerDetails

License issuer.

licenseArn :: Maybe Text

Amazon Resource Name (ARN) of the license.

licenseMetadata :: Maybe [Metadata]

License metadata.

createTime :: Maybe Text

License creation time.

newLicense :: License

Create a value of License with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:License', license_status - License status.

$sel:productName:License', license_productName - Product name.

$sel:licenseName:License', license_licenseName - License name.

$sel:productSKU:License', license_productSKU - Product SKU.

$sel:homeRegion:License', license_homeRegion - Home Region of the license.

$sel:version:License', license_version - License version.

$sel:entitlements:License', license_entitlements - License entitlements.

$sel:validity:License', license_validity - Date and time range during which the license is valid, in ISO8601-UTC format.

$sel:consumptionConfiguration:License', license_consumptionConfiguration - Configuration for consumption of the license.

$sel:beneficiary:License', license_beneficiary - License beneficiary.

$sel:issuer:License', license_issuer - License issuer.

$sel:licenseArn:License', license_licenseArn - Amazon Resource Name (ARN) of the license.

$sel:licenseMetadata:License', license_licenseMetadata - License metadata.

$sel:createTime:License', license_createTime - License creation time.

license_status :: Lens' License (Maybe LicenseStatus)

License status.

license_productName :: Lens' License (Maybe Text)

Product name.

license_licenseName :: Lens' License (Maybe Text)

License name.

license_productSKU :: Lens' License (Maybe Text)

Product SKU.

license_homeRegion :: Lens' License (Maybe Text)

Home Region of the license.

license_version :: Lens' License (Maybe Text)

License version.

license_entitlements :: Lens' License (Maybe [Entitlement])

License entitlements.

license_validity :: Lens' License (Maybe DatetimeRange)

Date and time range during which the license is valid, in ISO8601-UTC format.

license_consumptionConfiguration :: Lens' License (Maybe ConsumptionConfiguration)

Configuration for consumption of the license.

license_beneficiary :: Lens' License (Maybe Text)

License beneficiary.

license_issuer :: Lens' License (Maybe IssuerDetails)

License issuer.

license_licenseArn :: Lens' License (Maybe Text)

Amazon Resource Name (ARN) of the license.

license_licenseMetadata :: Lens' License (Maybe [Metadata])

License metadata.

license_createTime :: Lens' License (Maybe Text)

License creation time.