Network.AWS.LicenseManager.UpdateLicenseSpecificationsForResource

Description

Adds or removes the specified license configurations for the specified Amazon Web Services resource.

You can update the license specifications of AMIs, instances, and hosts. You cannot update the license specifications for launch templates and CloudFormation templates, as they send license configurations to the operation that creates the resource.

Creating a Request

data UpdateLicenseSpecificationsForResource

See: newUpdateLicenseSpecificationsForResource smart constructor.

Constructors

UpdateLicenseSpecificationsForResource'

Fields

addLicenseSpecifications :: Maybe [LicenseSpecification]

ARNs of the license configurations to add.

removeLicenseSpecifications :: Maybe [LicenseSpecification]

ARNs of the license configurations to remove.

resourceArn :: Text

Amazon Resource Name (ARN) of the Amazon Web Services resource.

newUpdateLicenseSpecificationsForResource

Arguments

:: Text
$sel:resourceArn:UpdateLicenseSpecificationsForResource'

-> UpdateLicenseSpecificationsForResource

Create a value of UpdateLicenseSpecificationsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:addLicenseSpecifications:UpdateLicenseSpecificationsForResource', updateLicenseSpecificationsForResource_addLicenseSpecifications - ARNs of the license configurations to add.

$sel:removeLicenseSpecifications:UpdateLicenseSpecificationsForResource', updateLicenseSpecificationsForResource_removeLicenseSpecifications - ARNs of the license configurations to remove.

$sel:resourceArn:UpdateLicenseSpecificationsForResource', updateLicenseSpecificationsForResource_resourceArn - Amazon Resource Name (ARN) of the Amazon Web Services resource.

Request Lenses

updateLicenseSpecificationsForResource_addLicenseSpecifications :: Lens' UpdateLicenseSpecificationsForResource (Maybe [LicenseSpecification])

ARNs of the license configurations to add.

updateLicenseSpecificationsForResource_removeLicenseSpecifications :: Lens' UpdateLicenseSpecificationsForResource (Maybe [LicenseSpecification])

ARNs of the license configurations to remove.

updateLicenseSpecificationsForResource_resourceArn :: Lens' UpdateLicenseSpecificationsForResource Text

Amazon Resource Name (ARN) of the Amazon Web Services resource.

Destructuring the Response

data UpdateLicenseSpecificationsForResourceResponse

See: newUpdateLicenseSpecificationsForResourceResponse smart constructor.

Constructors

UpdateLicenseSpecificationsForResourceResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateLicenseSpecificationsForResourceResponse

Arguments

:: Int
$sel:httpStatus:UpdateLicenseSpecificationsForResourceResponse'

-> UpdateLicenseSpecificationsForResourceResponse

Create a value of UpdateLicenseSpecificationsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateLicenseSpecificationsForResourceResponse', updateLicenseSpecificationsForResourceResponse_httpStatus - The response's http status code.

Response Lenses

updateLicenseSpecificationsForResourceResponse_httpStatus :: Lens' UpdateLicenseSpecificationsForResourceResponse Int

The response's http status code.