Network.AWS.Lightsail.CreateContactMethod

Description

Creates an email or SMS text message contact method.

A contact method is used to send you notifications about your Amazon Lightsail resources. You can add one email address and one mobile phone number contact method in each AWS Region. However, SMS text messaging is not supported in some AWS Regions, and SMS text messages cannot be sent to some countries/regions. For more information, see Notifications in Amazon Lightsail.

Creating a Request

data CreateContactMethod

See: newCreateContactMethod smart constructor.

Constructors

CreateContactMethod'

Fields

protocol :: ContactProtocol

The protocol of the contact method, such as Email or SMS (text messaging).

The SMS protocol is supported only in the following AWS Regions.

US East (N. Virginia) (us-east-1)

US West (Oregon) (us-west-2)

Europe (Ireland) (eu-west-1)

Asia Pacific (Tokyo) (ap-northeast-1)

Asia Pacific (Singapore) (ap-southeast-1)

Asia Pacific (Sydney) (ap-southeast-2)

For a list of countries/regions where SMS text messages can be sent, and the latest AWS Regions where SMS text messaging is supported, see Supported Regions and Countries in the Amazon SNS Developer Guide.

For more information about notifications in Amazon Lightsail, see Notifications in Amazon Lightsail.

contactEndpoint :: Text

The destination of the contact method, such as an email address or a mobile phone number.

Use the E.164 format when specifying a mobile phone number. E.164 is a standard for the phone number structure used for international telecommunication. Phone numbers that follow this format can have a maximum of 15 digits, and they are prefixed with the plus character (+) and the country code. For example, a U.S. phone number in E.164 format would be specified as +1XXX5550100. For more information, see E.164 on Wikipedia.

newCreateContactMethod

Arguments

:: ContactProtocol
$sel:protocol:CreateContactMethod'

-> Text
$sel:contactEndpoint:CreateContactMethod'

-> CreateContactMethod

Create a value of CreateContactMethod with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protocol:CreateContactMethod', createContactMethod_protocol - The protocol of the contact method, such as Email or SMS (text messaging).

The SMS protocol is supported only in the following AWS Regions.

US East (N. Virginia) (us-east-1)

US West (Oregon) (us-west-2)

Europe (Ireland) (eu-west-1)

Asia Pacific (Tokyo) (ap-northeast-1)

Asia Pacific (Singapore) (ap-southeast-1)

Asia Pacific (Sydney) (ap-southeast-2)

For a list of countries/regions where SMS text messages can be sent, and the latest AWS Regions where SMS text messaging is supported, see Supported Regions and Countries in the Amazon SNS Developer Guide.

For more information about notifications in Amazon Lightsail, see Notifications in Amazon Lightsail.

$sel:contactEndpoint:CreateContactMethod', createContactMethod_contactEndpoint - The destination of the contact method, such as an email address or a mobile phone number.

Use the E.164 format when specifying a mobile phone number. E.164 is a standard for the phone number structure used for international telecommunication. Phone numbers that follow this format can have a maximum of 15 digits, and they are prefixed with the plus character (+) and the country code. For example, a U.S. phone number in E.164 format would be specified as +1XXX5550100. For more information, see E.164 on Wikipedia.

Request Lenses

createContactMethod_protocol :: Lens' CreateContactMethod ContactProtocol

The protocol of the contact method, such as Email or SMS (text messaging).

The SMS protocol is supported only in the following AWS Regions.

US East (N. Virginia) (us-east-1)

US West (Oregon) (us-west-2)

Europe (Ireland) (eu-west-1)

Asia Pacific (Tokyo) (ap-northeast-1)

Asia Pacific (Singapore) (ap-southeast-1)

Asia Pacific (Sydney) (ap-southeast-2)

For a list of countries/regions where SMS text messages can be sent, and the latest AWS Regions where SMS text messaging is supported, see Supported Regions and Countries in the Amazon SNS Developer Guide.

For more information about notifications in Amazon Lightsail, see Notifications in Amazon Lightsail.

createContactMethod_contactEndpoint :: Lens' CreateContactMethod Text

The destination of the contact method, such as an email address or a mobile phone number.

Use the E.164 format when specifying a mobile phone number. E.164 is a standard for the phone number structure used for international telecommunication. Phone numbers that follow this format can have a maximum of 15 digits, and they are prefixed with the plus character (+) and the country code. For example, a U.S. phone number in E.164 format would be specified as +1XXX5550100. For more information, see E.164 on Wikipedia.

Destructuring the Response

data CreateContactMethodResponse

See: newCreateContactMethodResponse smart constructor.

Constructors

CreateContactMethodResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newCreateContactMethodResponse

Arguments

:: Int
$sel:httpStatus:CreateContactMethodResponse'

-> CreateContactMethodResponse

Create a value of CreateContactMethodResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:CreateContactMethodResponse', createContactMethodResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:CreateContactMethodResponse', createContactMethodResponse_httpStatus - The response's http status code.

Response Lenses

createContactMethodResponse_operations :: Lens' CreateContactMethodResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

createContactMethodResponse_httpStatus :: Lens' CreateContactMethodResponse Int

The response's http status code.