Network.AWS.Lightsail.CreateContainerService

Description

Creates an Amazon Lightsail container service.

A Lightsail container service is a compute resource to which you can deploy containers. For more information, see Container services in Amazon Lightsail in the Lightsail Dev Guide.

Creating a Request

data CreateContainerService

See: newCreateContainerService smart constructor.

Constructors

CreateContainerService'

Fields

tags :: Maybe [Tag]

The tag keys and optional values to add to the certificate during create.

Use the TagResource action to tag a resource after it's created.

For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

deployment :: Maybe ContainerServiceDeploymentRequest

An object that describes a deployment for the container service.

A deployment specifies the containers that will be launched on the container service and their settings, such as the ports to open, the environment variables to apply, and the launch command to run. It also specifies the container that will serve as the public endpoint of the deployment and its settings, such as the HTTP or HTTPS port to use, and the health check configuration.

publicDomainNames :: Maybe (HashMap Text [Text])

The public domain names to use with the container service, such as example.com and www.example.com.

You can specify up to four public domain names for a container service. The domain names that you specify are used when you create a deployment with a container configured as the public endpoint of your container service.

If you don't specify public domain names, then you can use the default domain of the container service.

You must create and validate an SSL/TLS certificate before you can use public domain names with your container service. Use the CreateCertificate action to create a certificate for the public domain names you want to use with your container service.

You can specify public domain names using a string to array map as shown in the example later on this page.

serviceName :: Text

The name for the container service.

The name that you specify for your container service will make up part of its default domain. The default domain of a container service is typically https://<ServiceName>.<RandomGUID>.<AWSRegion>.cs.amazonlightsail.com. If the name of your container service is container-service-1, and it's located in the US East (Ohio) AWS region (us-east-2), then the domain for your container service will be like the following example: https://container-service-1.ur4EXAMPLE2uq.us-east-2.cs.amazonlightsail.com

The following are the requirements for container service names:

Must be unique within each AWS Region in your Lightsail account.

Must contain 1 to 63 characters.

Must contain only alphanumeric characters and hyphens.

A hyphen (-) can separate words but cannot be at the start or end of the name.

power :: ContainerServicePowerName

The power specification for the container service.

The power specifies the amount of memory, vCPUs, and base monthly cost of each node of the container service. The power and scale of a container service makes up its configured capacity. To determine the monthly price of your container service, multiply the base price of the power with the scale (the number of nodes) of the service.

Use the GetContainerServicePowers action to get a list of power options that you can specify using this parameter, and their base monthly cost.

scale :: Natural

The scale specification for the container service.

The scale specifies the allocated compute nodes of the container service. The power and scale of a container service makes up its configured capacity. To determine the monthly price of your container service, multiply the base price of the power with the scale (the number of nodes) of the service.

newCreateContainerService

Arguments

:: Text
$sel:serviceName:CreateContainerService'

-> ContainerServicePowerName
$sel:power:CreateContainerService'

-> Natural
$sel:scale:CreateContainerService'

-> CreateContainerService

Create a value of CreateContainerService with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateContainerService', createContainerService_tags - The tag keys and optional values to add to the certificate during create.

Use the TagResource action to tag a resource after it's created.

For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

$sel:deployment:CreateContainerService', createContainerService_deployment - An object that describes a deployment for the container service.

A deployment specifies the containers that will be launched on the container service and their settings, such as the ports to open, the environment variables to apply, and the launch command to run. It also specifies the container that will serve as the public endpoint of the deployment and its settings, such as the HTTP or HTTPS port to use, and the health check configuration.

$sel:publicDomainNames:CreateContainerService', createContainerService_publicDomainNames - The public domain names to use with the container service, such as example.com and www.example.com.

You can specify up to four public domain names for a container service. The domain names that you specify are used when you create a deployment with a container configured as the public endpoint of your container service.

If you don't specify public domain names, then you can use the default domain of the container service.

You must create and validate an SSL/TLS certificate before you can use public domain names with your container service. Use the CreateCertificate action to create a certificate for the public domain names you want to use with your container service.

You can specify public domain names using a string to array map as shown in the example later on this page.

$sel:serviceName:CreateContainerService', createContainerService_serviceName - The name for the container service.

The name that you specify for your container service will make up part of its default domain. The default domain of a container service is typically https://<ServiceName>.<RandomGUID>.<AWSRegion>.cs.amazonlightsail.com. If the name of your container service is container-service-1, and it's located in the US East (Ohio) AWS region (us-east-2), then the domain for your container service will be like the following example: https://container-service-1.ur4EXAMPLE2uq.us-east-2.cs.amazonlightsail.com

The following are the requirements for container service names:

Must be unique within each AWS Region in your Lightsail account.

Must contain 1 to 63 characters.

Must contain only alphanumeric characters and hyphens.

A hyphen (-) can separate words but cannot be at the start or end of the name.

$sel:power:CreateContainerService', createContainerService_power - The power specification for the container service.

The power specifies the amount of memory, vCPUs, and base monthly cost of each node of the container service. The power and scale of a container service makes up its configured capacity. To determine the monthly price of your container service, multiply the base price of the power with the scale (the number of nodes) of the service.

Use the GetContainerServicePowers action to get a list of power options that you can specify using this parameter, and their base monthly cost.

$sel:scale:CreateContainerService', createContainerService_scale - The scale specification for the container service.

The scale specifies the allocated compute nodes of the container service. The power and scale of a container service makes up its configured capacity. To determine the monthly price of your container service, multiply the base price of the power with the scale (the number of nodes) of the service.

Request Lenses

createContainerService_tags :: Lens' CreateContainerService (Maybe [Tag])

The tag keys and optional values to add to the certificate during create.

Use the TagResource action to tag a resource after it's created.

For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

createContainerService_deployment :: Lens' CreateContainerService (Maybe ContainerServiceDeploymentRequest)

An object that describes a deployment for the container service.

A deployment specifies the containers that will be launched on the container service and their settings, such as the ports to open, the environment variables to apply, and the launch command to run. It also specifies the container that will serve as the public endpoint of the deployment and its settings, such as the HTTP or HTTPS port to use, and the health check configuration.

createContainerService_publicDomainNames :: Lens' CreateContainerService (Maybe (HashMap Text [Text]))

The public domain names to use with the container service, such as example.com and www.example.com.

You can specify up to four public domain names for a container service. The domain names that you specify are used when you create a deployment with a container configured as the public endpoint of your container service.

If you don't specify public domain names, then you can use the default domain of the container service.

You must create and validate an SSL/TLS certificate before you can use public domain names with your container service. Use the CreateCertificate action to create a certificate for the public domain names you want to use with your container service.

You can specify public domain names using a string to array map as shown in the example later on this page.

createContainerService_serviceName :: Lens' CreateContainerService Text

The name for the container service.

The name that you specify for your container service will make up part of its default domain. The default domain of a container service is typically https://<ServiceName>.<RandomGUID>.<AWSRegion>.cs.amazonlightsail.com. If the name of your container service is container-service-1, and it's located in the US East (Ohio) AWS region (us-east-2), then the domain for your container service will be like the following example: https://container-service-1.ur4EXAMPLE2uq.us-east-2.cs.amazonlightsail.com

The following are the requirements for container service names:

Must be unique within each AWS Region in your Lightsail account.

Must contain 1 to 63 characters.

Must contain only alphanumeric characters and hyphens.

A hyphen (-) can separate words but cannot be at the start or end of the name.

createContainerService_power :: Lens' CreateContainerService ContainerServicePowerName

The power specification for the container service.

The power specifies the amount of memory, vCPUs, and base monthly cost of each node of the container service. The power and scale of a container service makes up its configured capacity. To determine the monthly price of your container service, multiply the base price of the power with the scale (the number of nodes) of the service.

Use the GetContainerServicePowers action to get a list of power options that you can specify using this parameter, and their base monthly cost.

createContainerService_scale :: Lens' CreateContainerService Natural

The scale specification for the container service.

The scale specifies the allocated compute nodes of the container service. The power and scale of a container service makes up its configured capacity. To determine the monthly price of your container service, multiply the base price of the power with the scale (the number of nodes) of the service.

Destructuring the Response

data CreateContainerServiceResponse

See: newCreateContainerServiceResponse smart constructor.

Constructors

CreateContainerServiceResponse'

Fields

containerService :: Maybe ContainerService

An object that describes a container service.

httpStatus :: Int

The response's http status code.

newCreateContainerServiceResponse

Arguments

:: Int
$sel:httpStatus:CreateContainerServiceResponse'

-> CreateContainerServiceResponse

Create a value of CreateContainerServiceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerService:CreateContainerServiceResponse', createContainerServiceResponse_containerService - An object that describes a container service.

$sel:httpStatus:CreateContainerServiceResponse', createContainerServiceResponse_httpStatus - The response's http status code.

Response Lenses

createContainerServiceResponse_containerService :: Lens' CreateContainerServiceResponse (Maybe ContainerService)

An object that describes a container service.

createContainerServiceResponse_httpStatus :: Lens' CreateContainerServiceResponse Int

The response's http status code.