Network.AWS.Lightsail.CreateContainerServiceRegistryLogin

Description

Creates a temporary set of log in credentials that you can use to log in to the Docker process on your local machine. After you're logged in, you can use the native Docker commands to push your local container images to the container image registry of your Amazon Lightsail account so that you can use them with your Lightsail container service. The log in credentials expire 12 hours after they are created, at which point you will need to create a new set of log in credentials.

You can only push container images to the container service registry of your Lightsail account. You cannot pull container images or perform any other container image management actions on the container service registry.

After you push your container images to the container image registry of your Lightsail account, use the RegisterContainerImage action to register the pushed images to a specific Lightsail container service.

This action is not required if you install and use the Lightsail Control (lightsailctl) plugin to push container images to your Lightsail container service. For more information, see Pushing and managing container images on your Amazon Lightsail container services in the Amazon Lightsail Developer Guide.

Creating a Request

data CreateContainerServiceRegistryLogin

See: newCreateContainerServiceRegistryLogin smart constructor.

Constructors

CreateContainerServiceRegistryLogin'

newCreateContainerServiceRegistryLogin :: CreateContainerServiceRegistryLogin

Create a value of CreateContainerServiceRegistryLogin with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data CreateContainerServiceRegistryLoginResponse

See: newCreateContainerServiceRegistryLoginResponse smart constructor.

Constructors

CreateContainerServiceRegistryLoginResponse'

Fields

registryLogin :: Maybe ContainerServiceRegistryLogin

An object that describes the log in information for the container service registry of your Lightsail account.

httpStatus :: Int

The response's http status code.

newCreateContainerServiceRegistryLoginResponse

Arguments

:: Int
$sel:httpStatus:CreateContainerServiceRegistryLoginResponse'

-> CreateContainerServiceRegistryLoginResponse

Create a value of CreateContainerServiceRegistryLoginResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registryLogin:CreateContainerServiceRegistryLoginResponse', createContainerServiceRegistryLoginResponse_registryLogin - An object that describes the log in information for the container service registry of your Lightsail account.

$sel:httpStatus:CreateContainerServiceRegistryLoginResponse', createContainerServiceRegistryLoginResponse_httpStatus - The response's http status code.

Response Lenses

createContainerServiceRegistryLoginResponse_registryLogin :: Lens' CreateContainerServiceRegistryLoginResponse (Maybe ContainerServiceRegistryLogin)

An object that describes the log in information for the container service registry of your Lightsail account.

createContainerServiceRegistryLoginResponse_httpStatus :: Lens' CreateContainerServiceRegistryLoginResponse Int

The response's http status code.