Network.AWS.Lightsail.CreateDiskSnapshot

Description

Creates a snapshot of a block storage disk. You can use snapshots for backups, to make copies of disks, and to save data before shutting down a Lightsail instance.

You can take a snapshot of an attached disk that is in use; however, snapshots only capture data that has been written to your disk at the time the snapshot command is issued. This may exclude any data that has been cached by any applications or the operating system. If you can pause any file systems on the disk long enough to take a snapshot, your snapshot should be complete. Nevertheless, if you cannot pause all file writes to the disk, you should unmount the disk from within the Lightsail instance, issue the create disk snapshot command, and then remount the disk to ensure a consistent and complete snapshot. You may remount and use your disk while the snapshot status is pending.

You can also use this operation to create a snapshot of an instance's system volume. You might want to do this, for example, to recover data from the system volume of a botched instance or to create a backup of the system volume like you would for a block storage disk. To create a snapshot of a system volume, just define the instance name parameter when issuing the snapshot command, and a snapshot of the defined instance's system volume will be created. After the snapshot is available, you can create a block storage disk from the snapshot and attach it to a running instance to access the data on the disk.

The create disk snapshot operation supports tag-based access control via request tags. For more information, see the Amazon Lightsail Developer Guide.

Creating a Request

data CreateDiskSnapshot

See: newCreateDiskSnapshot smart constructor.

Constructors

CreateDiskSnapshot'

Fields

instanceName :: Maybe Text

The unique name of the source instance (e.g., Amazon_Linux-512MB-Virginia-1). When this is defined, a snapshot of the instance's system volume is created.

This parameter cannot be defined together with the disk name parameter. The instance name and disk name parameters are mutually exclusive.

tags :: Maybe [Tag]

The tag keys and optional values to add to the resource during create.

Use the TagResource action to tag a resource after it's created.

diskName :: Maybe Text

The unique name of the source disk (e.g., Disk-Virginia-1).

This parameter cannot be defined together with the instance name parameter. The disk name and instance name parameters are mutually exclusive.

diskSnapshotName :: Text

The name of the destination disk snapshot (e.g., my-disk-snapshot) based on the source disk.

newCreateDiskSnapshot

Arguments

:: Text
$sel:diskSnapshotName:CreateDiskSnapshot'

-> CreateDiskSnapshot

Create a value of CreateDiskSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceName:CreateDiskSnapshot', createDiskSnapshot_instanceName - The unique name of the source instance (e.g., Amazon_Linux-512MB-Virginia-1). When this is defined, a snapshot of the instance's system volume is created.

This parameter cannot be defined together with the disk name parameter. The instance name and disk name parameters are mutually exclusive.

$sel:tags:CreateDiskSnapshot', createDiskSnapshot_tags - The tag keys and optional values to add to the resource during create.

Use the TagResource action to tag a resource after it's created.

$sel:diskName:CreateDiskSnapshot', createDiskSnapshot_diskName - The unique name of the source disk (e.g., Disk-Virginia-1).

This parameter cannot be defined together with the instance name parameter. The disk name and instance name parameters are mutually exclusive.

$sel:diskSnapshotName:CreateDiskSnapshot', createDiskSnapshot_diskSnapshotName - The name of the destination disk snapshot (e.g., my-disk-snapshot) based on the source disk.

Request Lenses

createDiskSnapshot_instanceName :: Lens' CreateDiskSnapshot (Maybe Text)

The unique name of the source instance (e.g., Amazon_Linux-512MB-Virginia-1). When this is defined, a snapshot of the instance's system volume is created.

This parameter cannot be defined together with the disk name parameter. The instance name and disk name parameters are mutually exclusive.

createDiskSnapshot_tags :: Lens' CreateDiskSnapshot (Maybe [Tag])

The tag keys and optional values to add to the resource during create.

Use the TagResource action to tag a resource after it's created.

createDiskSnapshot_diskName :: Lens' CreateDiskSnapshot (Maybe Text)

The unique name of the source disk (e.g., Disk-Virginia-1).

This parameter cannot be defined together with the instance name parameter. The disk name and instance name parameters are mutually exclusive.

createDiskSnapshot_diskSnapshotName :: Lens' CreateDiskSnapshot Text

The name of the destination disk snapshot (e.g., my-disk-snapshot) based on the source disk.

Destructuring the Response

data CreateDiskSnapshotResponse

See: newCreateDiskSnapshotResponse smart constructor.

Constructors

CreateDiskSnapshotResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newCreateDiskSnapshotResponse

Arguments

:: Int
$sel:httpStatus:CreateDiskSnapshotResponse'

-> CreateDiskSnapshotResponse

Create a value of CreateDiskSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:CreateDiskSnapshotResponse', createDiskSnapshotResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:CreateDiskSnapshotResponse', createDiskSnapshotResponse_httpStatus - The response's http status code.

Response Lenses

createDiskSnapshotResponse_operations :: Lens' CreateDiskSnapshotResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

createDiskSnapshotResponse_httpStatus :: Lens' CreateDiskSnapshotResponse Int

The response's http status code.