Network.AWS.Lightsail.CreateDistribution

Description

Creates an Amazon Lightsail content delivery network (CDN) distribution.

A distribution is a globally distributed network of caching servers that improve the performance of your website or web application hosted on a Lightsail instance. For more information, see Content delivery networks in Amazon Lightsail.

Creating a Request

data CreateDistribution

See: newCreateDistribution smart constructor.

Constructors

CreateDistribution'

Fields

ipAddressType :: Maybe IpAddressType

The IP address type for the distribution.

The possible values are ipv4 for IPv4 only, and dualstack for IPv4 and IPv6.

The default value is dualstack.

cacheBehaviorSettings :: Maybe CacheSettings

An object that describes the cache behavior settings for the distribution.

tags :: Maybe [Tag]

The tag keys and optional values to add to the distribution during create.

Use the TagResource action to tag a resource after it's created.

cacheBehaviors :: Maybe [CacheBehaviorPerPath]

An array of objects that describe the per-path cache behavior for the distribution.

distributionName :: Text

The name for the distribution.

origin :: InputOrigin

An object that describes the origin resource for the distribution, such as a Lightsail instance or load balancer.

The distribution pulls, caches, and serves content from the origin.

defaultCacheBehavior :: CacheBehavior

An object that describes the default cache behavior for the distribution.

bundleId :: Text

The bundle ID to use for the distribution.

A distribution bundle describes the specifications of your distribution, such as the monthly cost and monthly network transfer quota.

Use the GetDistributionBundles action to get a list of distribution bundle IDs that you can specify.

newCreateDistribution

Arguments

:: Text
$sel:distributionName:CreateDistribution'

-> InputOrigin
$sel:origin:CreateDistribution'

-> CacheBehavior
$sel:defaultCacheBehavior:CreateDistribution'

-> Text
$sel:bundleId:CreateDistribution'

-> CreateDistribution

Create a value of CreateDistribution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipAddressType:CreateDistribution', createDistribution_ipAddressType - The IP address type for the distribution.

The possible values are ipv4 for IPv4 only, and dualstack for IPv4 and IPv6.

The default value is dualstack.

$sel:cacheBehaviorSettings:CreateDistribution', createDistribution_cacheBehaviorSettings - An object that describes the cache behavior settings for the distribution.

$sel:tags:CreateDistribution', createDistribution_tags - The tag keys and optional values to add to the distribution during create.

Use the TagResource action to tag a resource after it's created.

$sel:cacheBehaviors:CreateDistribution', createDistribution_cacheBehaviors - An array of objects that describe the per-path cache behavior for the distribution.

$sel:distributionName:CreateDistribution', createDistribution_distributionName - The name for the distribution.

$sel:origin:CreateDistribution', createDistribution_origin - An object that describes the origin resource for the distribution, such as a Lightsail instance or load balancer.

The distribution pulls, caches, and serves content from the origin.

$sel:defaultCacheBehavior:CreateDistribution', createDistribution_defaultCacheBehavior - An object that describes the default cache behavior for the distribution.

$sel:bundleId:CreateDistribution', createDistribution_bundleId - The bundle ID to use for the distribution.

A distribution bundle describes the specifications of your distribution, such as the monthly cost and monthly network transfer quota.

Use the GetDistributionBundles action to get a list of distribution bundle IDs that you can specify.

Request Lenses

createDistribution_ipAddressType :: Lens' CreateDistribution (Maybe IpAddressType)

The IP address type for the distribution.

The possible values are ipv4 for IPv4 only, and dualstack for IPv4 and IPv6.

The default value is dualstack.

createDistribution_cacheBehaviorSettings :: Lens' CreateDistribution (Maybe CacheSettings)

An object that describes the cache behavior settings for the distribution.

createDistribution_tags :: Lens' CreateDistribution (Maybe [Tag])

The tag keys and optional values to add to the distribution during create.

Use the TagResource action to tag a resource after it's created.

createDistribution_cacheBehaviors :: Lens' CreateDistribution (Maybe [CacheBehaviorPerPath])

An array of objects that describe the per-path cache behavior for the distribution.

createDistribution_distributionName :: Lens' CreateDistribution Text

The name for the distribution.

createDistribution_origin :: Lens' CreateDistribution InputOrigin

An object that describes the origin resource for the distribution, such as a Lightsail instance or load balancer.

The distribution pulls, caches, and serves content from the origin.

createDistribution_defaultCacheBehavior :: Lens' CreateDistribution CacheBehavior

An object that describes the default cache behavior for the distribution.

createDistribution_bundleId :: Lens' CreateDistribution Text

The bundle ID to use for the distribution.

A distribution bundle describes the specifications of your distribution, such as the monthly cost and monthly network transfer quota.

Use the GetDistributionBundles action to get a list of distribution bundle IDs that you can specify.

Destructuring the Response

data CreateDistributionResponse

See: newCreateDistributionResponse smart constructor.

Constructors

CreateDistributionResponse'

Fields

operation :: Maybe Operation

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

distribution :: Maybe LightsailDistribution

An object that describes the distribution created.

httpStatus :: Int

The response's http status code.

newCreateDistributionResponse

Arguments

:: Int
$sel:httpStatus:CreateDistributionResponse'

-> CreateDistributionResponse

Create a value of CreateDistributionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operation:CreateDistributionResponse', createDistributionResponse_operation - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:distribution:CreateDistributionResponse', createDistributionResponse_distribution - An object that describes the distribution created.

$sel:httpStatus:CreateDistributionResponse', createDistributionResponse_httpStatus - The response's http status code.

Response Lenses

createDistributionResponse_operation :: Lens' CreateDistributionResponse (Maybe Operation)

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

createDistributionResponse_distribution :: Lens' CreateDistributionResponse (Maybe LightsailDistribution)

An object that describes the distribution created.

createDistributionResponse_httpStatus :: Lens' CreateDistributionResponse Int

The response's http status code.