Network.AWS.Lightsail.CreateDomainEntry

Description

Creates one of the following domain name system (DNS) records in a domain DNS zone: Address (A), canonical name (CNAME), mail exchanger (MX), name server (NS), start of authority (SOA), service locator (SRV), or text (TXT).

The create domain entry operation supports tag-based access control via resource tags applied to the resource identified by domain name. For more information, see the Amazon Lightsail Developer Guide.

Creating a Request

data CreateDomainEntry

See: newCreateDomainEntry smart constructor.

Constructors

CreateDomainEntry'

Fields

domainName :: Text

The domain name (e.g., example.com) for which you want to create the domain entry.

domainEntry :: DomainEntry

An array of key-value pairs containing information about the domain entry request.

newCreateDomainEntry

Arguments

:: Text
$sel:domainName:CreateDomainEntry'

-> DomainEntry
$sel:domainEntry:CreateDomainEntry'

-> CreateDomainEntry

Create a value of CreateDomainEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:domainName:CreateDomainEntry', createDomainEntry_domainName - The domain name (e.g., example.com) for which you want to create the domain entry.

$sel:domainEntry:CreateDomainEntry', createDomainEntry_domainEntry - An array of key-value pairs containing information about the domain entry request.

Request Lenses

createDomainEntry_domainName :: Lens' CreateDomainEntry Text

The domain name (e.g., example.com) for which you want to create the domain entry.

createDomainEntry_domainEntry :: Lens' CreateDomainEntry DomainEntry

An array of key-value pairs containing information about the domain entry request.

Destructuring the Response

data CreateDomainEntryResponse

See: newCreateDomainEntryResponse smart constructor.

Constructors

CreateDomainEntryResponse'

Fields

operation :: Maybe Operation

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newCreateDomainEntryResponse

Arguments

:: Int
$sel:httpStatus:CreateDomainEntryResponse'

-> CreateDomainEntryResponse

Create a value of CreateDomainEntryResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operation:CreateDomainEntryResponse', createDomainEntryResponse_operation - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:CreateDomainEntryResponse', createDomainEntryResponse_httpStatus - The response's http status code.

Response Lenses

createDomainEntryResponse_operation :: Lens' CreateDomainEntryResponse (Maybe Operation)

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

createDomainEntryResponse_httpStatus :: Lens' CreateDomainEntryResponse Int

The response's http status code.