Network.AWS.Lightsail.CreateInstanceSnapshot

Description

Creates a snapshot of a specific virtual private server, or instance. You can use a snapshot to create a new instance that is based on that snapshot.

The create instance snapshot operation supports tag-based access control via request tags. For more information, see the Amazon Lightsail Developer Guide.

Creating a Request

data CreateInstanceSnapshot

See: newCreateInstanceSnapshot smart constructor.

Constructors

CreateInstanceSnapshot'

Fields

tags :: Maybe [Tag]

The tag keys and optional values to add to the resource during create.

Use the TagResource action to tag a resource after it's created.

instanceSnapshotName :: Text

The name for your new snapshot.

instanceName :: Text

The Lightsail instance on which to base your snapshot.

newCreateInstanceSnapshot

Arguments

:: Text
$sel:instanceSnapshotName:CreateInstanceSnapshot'

-> Text
$sel:instanceName:CreateInstanceSnapshot'

-> CreateInstanceSnapshot

Create a value of CreateInstanceSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateInstanceSnapshot', createInstanceSnapshot_tags - The tag keys and optional values to add to the resource during create.

Use the TagResource action to tag a resource after it's created.

$sel:instanceSnapshotName:CreateInstanceSnapshot', createInstanceSnapshot_instanceSnapshotName - The name for your new snapshot.

$sel:instanceName:CreateInstanceSnapshot', createInstanceSnapshot_instanceName - The Lightsail instance on which to base your snapshot.

Request Lenses

createInstanceSnapshot_tags :: Lens' CreateInstanceSnapshot (Maybe [Tag])

The tag keys and optional values to add to the resource during create.

Use the TagResource action to tag a resource after it's created.

createInstanceSnapshot_instanceSnapshotName :: Lens' CreateInstanceSnapshot Text

The name for your new snapshot.

createInstanceSnapshot_instanceName :: Lens' CreateInstanceSnapshot Text

The Lightsail instance on which to base your snapshot.

Destructuring the Response

data CreateInstanceSnapshotResponse

See: newCreateInstanceSnapshotResponse smart constructor.

Constructors

CreateInstanceSnapshotResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newCreateInstanceSnapshotResponse

Arguments

:: Int
$sel:httpStatus:CreateInstanceSnapshotResponse'

-> CreateInstanceSnapshotResponse

Create a value of CreateInstanceSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:CreateInstanceSnapshotResponse', createInstanceSnapshotResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:CreateInstanceSnapshotResponse', createInstanceSnapshotResponse_httpStatus - The response's http status code.

Response Lenses

createInstanceSnapshotResponse_operations :: Lens' CreateInstanceSnapshotResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

createInstanceSnapshotResponse_httpStatus :: Lens' CreateInstanceSnapshotResponse Int

The response's http status code.