Network.AWS.Lightsail.CreateInstancesFromSnapshot

Description

Creates one or more new instances from a manual or automatic snapshot of an instance.

The create instances from snapshot operation supports tag-based access control via request tags and resource tags applied to the resource identified by instance snapshot name. For more information, see the Amazon Lightsail Developer Guide.

Creating a Request

data CreateInstancesFromSnapshot

See: newCreateInstancesFromSnapshot smart constructor.

Constructors

CreateInstancesFromSnapshot'

Fields

ipAddressType :: Maybe IpAddressType

The IP address type for the instance.

The possible values are ipv4 for IPv4 only, and dualstack for IPv4 and IPv6.

The default value is dualstack.

restoreDate :: Maybe Text

The date of the automatic snapshot to use for the new instance. Use the get auto snapshots operation to identify the dates of the available automatic snapshots.

Constraints:

Must be specified in YYYY-MM-DD format.

This parameter cannot be defined together with the use latest restorable auto snapshot parameter. The restore date and use latest restorable auto snapshot parameters are mutually exclusive.

Define this parameter only when creating a new instance from an automatic snapshot. For more information, see the Amazon Lightsail Developer Guide.

userData :: Maybe Text

You can create a launch script that configures a server with additional user data. For example, apt-get -y update.

Depending on the machine image you choose, the command to get software on your instance varies. Amazon Linux and CentOS use yum, Debian and Ubuntu use apt-get, and FreeBSD uses pkg. For a complete list, see the Amazon Lightsail Developer Guide.

addOns :: Maybe [AddOnRequest]

An array of objects representing the add-ons to enable for the new instance.

attachedDiskMapping :: Maybe (HashMap Text [DiskMap])

An object containing information about one or more disk mappings.

instanceSnapshotName :: Maybe Text

The name of the instance snapshot on which you are basing your new instances. Use the get instance snapshots operation to return information about your existing snapshots.

Constraint:

This parameter cannot be defined together with the source instance name parameter. The instance snapshot name and source instance name parameters are mutually exclusive.

keyPairName :: Maybe Text

The name for your key pair.

useLatestRestorableAutoSnapshot :: Maybe Bool

A Boolean value to indicate whether to use the latest available automatic snapshot.

Constraints:

This parameter cannot be defined together with the restore date parameter. The use latest restorable auto snapshot and restore date parameters are mutually exclusive.

Define this parameter only when creating a new instance from an automatic snapshot. For more information, see the Amazon Lightsail Developer Guide.

tags :: Maybe [Tag]

The tag keys and optional values to add to the resource during create.

Use the TagResource action to tag a resource after it's created.

sourceInstanceName :: Maybe Text

The name of the source instance from which the source automatic snapshot was created.

Constraints:

This parameter cannot be defined together with the instance snapshot name parameter. The source instance name and instance snapshot name parameters are mutually exclusive.

Define this parameter only when creating a new instance from an automatic snapshot. For more information, see the Amazon Lightsail Developer Guide.

instanceNames :: [Text]

The names for your new instances.

availabilityZone :: Text

The Availability Zone where you want to create your instances. Use the following formatting: us-east-2a (case sensitive). You can get a list of Availability Zones by using the get regions operation. Be sure to add the include Availability Zones parameter to your request.

bundleId :: Text

The bundle of specification information for your virtual private server (or instance), including the pricing plan (e.g., micro_1_0).

newCreateInstancesFromSnapshot

Arguments

:: Text
$sel:availabilityZone:CreateInstancesFromSnapshot'

-> Text
$sel:bundleId:CreateInstancesFromSnapshot'

-> CreateInstancesFromSnapshot

Create a value of CreateInstancesFromSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipAddressType:CreateInstancesFromSnapshot', createInstancesFromSnapshot_ipAddressType - The IP address type for the instance.

The possible values are ipv4 for IPv4 only, and dualstack for IPv4 and IPv6.

The default value is dualstack.

$sel:restoreDate:CreateInstancesFromSnapshot', createInstancesFromSnapshot_restoreDate - The date of the automatic snapshot to use for the new instance. Use the get auto snapshots operation to identify the dates of the available automatic snapshots.

Constraints:

Must be specified in YYYY-MM-DD format.

This parameter cannot be defined together with the use latest restorable auto snapshot parameter. The restore date and use latest restorable auto snapshot parameters are mutually exclusive.

Define this parameter only when creating a new instance from an automatic snapshot. For more information, see the Amazon Lightsail Developer Guide.

$sel:userData:CreateInstancesFromSnapshot', createInstancesFromSnapshot_userData - You can create a launch script that configures a server with additional user data. For example, apt-get -y update.

Depending on the machine image you choose, the command to get software on your instance varies. Amazon Linux and CentOS use yum, Debian and Ubuntu use apt-get, and FreeBSD uses pkg. For a complete list, see the Amazon Lightsail Developer Guide.

$sel:addOns:CreateInstancesFromSnapshot', createInstancesFromSnapshot_addOns - An array of objects representing the add-ons to enable for the new instance.

$sel:attachedDiskMapping:CreateInstancesFromSnapshot', createInstancesFromSnapshot_attachedDiskMapping - An object containing information about one or more disk mappings.

$sel:instanceSnapshotName:CreateInstancesFromSnapshot', createInstancesFromSnapshot_instanceSnapshotName - The name of the instance snapshot on which you are basing your new instances. Use the get instance snapshots operation to return information about your existing snapshots.

Constraint:

This parameter cannot be defined together with the source instance name parameter. The instance snapshot name and source instance name parameters are mutually exclusive.

$sel:keyPairName:CreateInstancesFromSnapshot', createInstancesFromSnapshot_keyPairName - The name for your key pair.

$sel:useLatestRestorableAutoSnapshot:CreateInstancesFromSnapshot', createInstancesFromSnapshot_useLatestRestorableAutoSnapshot - A Boolean value to indicate whether to use the latest available automatic snapshot.

Constraints:

This parameter cannot be defined together with the restore date parameter. The use latest restorable auto snapshot and restore date parameters are mutually exclusive.

Define this parameter only when creating a new instance from an automatic snapshot. For more information, see the Amazon Lightsail Developer Guide.

$sel:tags:CreateInstancesFromSnapshot', createInstancesFromSnapshot_tags - The tag keys and optional values to add to the resource during create.

Use the TagResource action to tag a resource after it's created.

$sel:sourceInstanceName:CreateInstancesFromSnapshot', createInstancesFromSnapshot_sourceInstanceName - The name of the source instance from which the source automatic snapshot was created.

Constraints:

This parameter cannot be defined together with the instance snapshot name parameter. The source instance name and instance snapshot name parameters are mutually exclusive.

Define this parameter only when creating a new instance from an automatic snapshot. For more information, see the Amazon Lightsail Developer Guide.

$sel:instanceNames:CreateInstancesFromSnapshot', createInstancesFromSnapshot_instanceNames - The names for your new instances.

$sel:availabilityZone:CreateInstancesFromSnapshot', createInstancesFromSnapshot_availabilityZone - The Availability Zone where you want to create your instances. Use the following formatting: us-east-2a (case sensitive). You can get a list of Availability Zones by using the get regions operation. Be sure to add the include Availability Zones parameter to your request.

$sel:bundleId:CreateInstancesFromSnapshot', createInstancesFromSnapshot_bundleId - The bundle of specification information for your virtual private server (or instance), including the pricing plan (e.g., micro_1_0).

Request Lenses

createInstancesFromSnapshot_ipAddressType :: Lens' CreateInstancesFromSnapshot (Maybe IpAddressType)

The IP address type for the instance.

The possible values are ipv4 for IPv4 only, and dualstack for IPv4 and IPv6.

The default value is dualstack.

createInstancesFromSnapshot_restoreDate :: Lens' CreateInstancesFromSnapshot (Maybe Text)

The date of the automatic snapshot to use for the new instance. Use the get auto snapshots operation to identify the dates of the available automatic snapshots.

Constraints:

Must be specified in YYYY-MM-DD format.

This parameter cannot be defined together with the use latest restorable auto snapshot parameter. The restore date and use latest restorable auto snapshot parameters are mutually exclusive.

Define this parameter only when creating a new instance from an automatic snapshot. For more information, see the Amazon Lightsail Developer Guide.

createInstancesFromSnapshot_userData :: Lens' CreateInstancesFromSnapshot (Maybe Text)

You can create a launch script that configures a server with additional user data. For example, apt-get -y update.

Depending on the machine image you choose, the command to get software on your instance varies. Amazon Linux and CentOS use yum, Debian and Ubuntu use apt-get, and FreeBSD uses pkg. For a complete list, see the Amazon Lightsail Developer Guide.

createInstancesFromSnapshot_addOns :: Lens' CreateInstancesFromSnapshot (Maybe [AddOnRequest])

An array of objects representing the add-ons to enable for the new instance.

createInstancesFromSnapshot_attachedDiskMapping :: Lens' CreateInstancesFromSnapshot (Maybe (HashMap Text [DiskMap]))

An object containing information about one or more disk mappings.

createInstancesFromSnapshot_instanceSnapshotName :: Lens' CreateInstancesFromSnapshot (Maybe Text)

The name of the instance snapshot on which you are basing your new instances. Use the get instance snapshots operation to return information about your existing snapshots.

Constraint:

This parameter cannot be defined together with the source instance name parameter. The instance snapshot name and source instance name parameters are mutually exclusive.

createInstancesFromSnapshot_keyPairName :: Lens' CreateInstancesFromSnapshot (Maybe Text)

The name for your key pair.

createInstancesFromSnapshot_useLatestRestorableAutoSnapshot :: Lens' CreateInstancesFromSnapshot (Maybe Bool)

A Boolean value to indicate whether to use the latest available automatic snapshot.

Constraints:

This parameter cannot be defined together with the restore date parameter. The use latest restorable auto snapshot and restore date parameters are mutually exclusive.

Define this parameter only when creating a new instance from an automatic snapshot. For more information, see the Amazon Lightsail Developer Guide.

createInstancesFromSnapshot_tags :: Lens' CreateInstancesFromSnapshot (Maybe [Tag])

The tag keys and optional values to add to the resource during create.

Use the TagResource action to tag a resource after it's created.

createInstancesFromSnapshot_sourceInstanceName :: Lens' CreateInstancesFromSnapshot (Maybe Text)

The name of the source instance from which the source automatic snapshot was created.

Constraints:

This parameter cannot be defined together with the instance snapshot name parameter. The source instance name and instance snapshot name parameters are mutually exclusive.

Define this parameter only when creating a new instance from an automatic snapshot. For more information, see the Amazon Lightsail Developer Guide.

createInstancesFromSnapshot_instanceNames :: Lens' CreateInstancesFromSnapshot [Text]

The names for your new instances.

createInstancesFromSnapshot_availabilityZone :: Lens' CreateInstancesFromSnapshot Text

The Availability Zone where you want to create your instances. Use the following formatting: us-east-2a (case sensitive). You can get a list of Availability Zones by using the get regions operation. Be sure to add the include Availability Zones parameter to your request.

createInstancesFromSnapshot_bundleId :: Lens' CreateInstancesFromSnapshot Text

The bundle of specification information for your virtual private server (or instance), including the pricing plan (e.g., micro_1_0).

Destructuring the Response

data CreateInstancesFromSnapshotResponse

See: newCreateInstancesFromSnapshotResponse smart constructor.

Constructors

CreateInstancesFromSnapshotResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newCreateInstancesFromSnapshotResponse

Arguments

:: Int
$sel:httpStatus:CreateInstancesFromSnapshotResponse'

-> CreateInstancesFromSnapshotResponse

Create a value of CreateInstancesFromSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:CreateInstancesFromSnapshotResponse', createInstancesFromSnapshotResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:CreateInstancesFromSnapshotResponse', createInstancesFromSnapshotResponse_httpStatus - The response's http status code.

Response Lenses

createInstancesFromSnapshotResponse_operations :: Lens' CreateInstancesFromSnapshotResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

createInstancesFromSnapshotResponse_httpStatus :: Lens' CreateInstancesFromSnapshotResponse Int

The response's http status code.