Network.AWS.Lightsail.CreateLoadBalancer

Description

Creates a Lightsail load balancer. To learn more about deciding whether to load balance your application, see Configure your Lightsail instances for load balancing. You can create up to 5 load balancers per AWS Region in your account.

When you create a load balancer, you can specify a unique name and port settings. To change additional load balancer settings, use the UpdateLoadBalancerAttribute operation.

The create load balancer operation supports tag-based access control via request tags. For more information, see the Amazon Lightsail Developer Guide.

Creating a Request

data CreateLoadBalancer

See: newCreateLoadBalancer smart constructor.

Constructors

CreateLoadBalancer'

Fields

healthCheckPath :: Maybe Text

The path you provided to perform the load balancer health check. If you didn't specify a health check path, Lightsail uses the root path of your website (e.g., "/").

You may want to specify a custom health check path other than the root of your application if your home page loads slowly or has a lot of media or scripting on it.

certificateName :: Maybe Text

The name of the SSL/TLS certificate.

If you specify certificateName, then certificateDomainName is required (and vice-versa).

certificateDomainName :: Maybe Text

The domain name with which your certificate is associated (e.g., example.com).

If you specify certificateDomainName, then certificateName is required (and vice-versa).

certificateAlternativeNames :: Maybe [Text]

The optional alternative domains and subdomains to use with your SSL/TLS certificate (e.g., www.example.com, example.com, m.example.com, blog.example.com).

ipAddressType :: Maybe IpAddressType

The IP address type for the load balancer.

The possible values are ipv4 for IPv4 only, and dualstack for IPv4 and IPv6.

The default value is dualstack.

tags :: Maybe [Tag]

The tag keys and optional values to add to the resource during create.

Use the TagResource action to tag a resource after it's created.

loadBalancerName :: Text

The name of your load balancer.

instancePort :: Int

The instance port where you're creating your load balancer.

newCreateLoadBalancer

Arguments

:: Text
$sel:loadBalancerName:CreateLoadBalancer'

-> Int
$sel:instancePort:CreateLoadBalancer'

-> CreateLoadBalancer

Create a value of CreateLoadBalancer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:healthCheckPath:CreateLoadBalancer', createLoadBalancer_healthCheckPath - The path you provided to perform the load balancer health check. If you didn't specify a health check path, Lightsail uses the root path of your website (e.g., "/").

You may want to specify a custom health check path other than the root of your application if your home page loads slowly or has a lot of media or scripting on it.

$sel:certificateName:CreateLoadBalancer', createLoadBalancer_certificateName - The name of the SSL/TLS certificate.

If you specify certificateName, then certificateDomainName is required (and vice-versa).

$sel:certificateDomainName:CreateLoadBalancer', createLoadBalancer_certificateDomainName - The domain name with which your certificate is associated (e.g., example.com).

If you specify certificateDomainName, then certificateName is required (and vice-versa).

$sel:certificateAlternativeNames:CreateLoadBalancer', createLoadBalancer_certificateAlternativeNames - The optional alternative domains and subdomains to use with your SSL/TLS certificate (e.g., www.example.com, example.com, m.example.com, blog.example.com).

$sel:ipAddressType:CreateLoadBalancer', createLoadBalancer_ipAddressType - The IP address type for the load balancer.

The possible values are ipv4 for IPv4 only, and dualstack for IPv4 and IPv6.

The default value is dualstack.

$sel:tags:CreateLoadBalancer', createLoadBalancer_tags - The tag keys and optional values to add to the resource during create.

Use the TagResource action to tag a resource after it's created.

$sel:loadBalancerName:CreateLoadBalancer', createLoadBalancer_loadBalancerName - The name of your load balancer.

$sel:instancePort:CreateLoadBalancer', createLoadBalancer_instancePort - The instance port where you're creating your load balancer.

Request Lenses

createLoadBalancer_healthCheckPath :: Lens' CreateLoadBalancer (Maybe Text)

The path you provided to perform the load balancer health check. If you didn't specify a health check path, Lightsail uses the root path of your website (e.g., "/").

You may want to specify a custom health check path other than the root of your application if your home page loads slowly or has a lot of media or scripting on it.

createLoadBalancer_certificateName :: Lens' CreateLoadBalancer (Maybe Text)

The name of the SSL/TLS certificate.

If you specify certificateName, then certificateDomainName is required (and vice-versa).

createLoadBalancer_certificateDomainName :: Lens' CreateLoadBalancer (Maybe Text)

The domain name with which your certificate is associated (e.g., example.com).

If you specify certificateDomainName, then certificateName is required (and vice-versa).

createLoadBalancer_certificateAlternativeNames :: Lens' CreateLoadBalancer (Maybe [Text])

The optional alternative domains and subdomains to use with your SSL/TLS certificate (e.g., www.example.com, example.com, m.example.com, blog.example.com).

createLoadBalancer_ipAddressType :: Lens' CreateLoadBalancer (Maybe IpAddressType)

The IP address type for the load balancer.

The possible values are ipv4 for IPv4 only, and dualstack for IPv4 and IPv6.

The default value is dualstack.

createLoadBalancer_tags :: Lens' CreateLoadBalancer (Maybe [Tag])

The tag keys and optional values to add to the resource during create.

Use the TagResource action to tag a resource after it's created.

createLoadBalancer_loadBalancerName :: Lens' CreateLoadBalancer Text

The name of your load balancer.

createLoadBalancer_instancePort :: Lens' CreateLoadBalancer Int

The instance port where you're creating your load balancer.

Destructuring the Response

data CreateLoadBalancerResponse

See: newCreateLoadBalancerResponse smart constructor.

Constructors

CreateLoadBalancerResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newCreateLoadBalancerResponse

Arguments

:: Int
$sel:httpStatus:CreateLoadBalancerResponse'

-> CreateLoadBalancerResponse

Create a value of CreateLoadBalancerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:CreateLoadBalancerResponse', createLoadBalancerResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:CreateLoadBalancerResponse', createLoadBalancerResponse_httpStatus - The response's http status code.

Response Lenses

createLoadBalancerResponse_operations :: Lens' CreateLoadBalancerResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

createLoadBalancerResponse_httpStatus :: Lens' CreateLoadBalancerResponse Int

The response's http status code.