Network.AWS.Lightsail.CreateRelationalDatabase

Description

Creates a new database in Amazon Lightsail.

The create relational database operation supports tag-based access control via request tags. For more information, see the Amazon Lightsail Developer Guide.

Creating a Request

data CreateRelationalDatabase

See: newCreateRelationalDatabase smart constructor.

Constructors

CreateRelationalDatabase'

Fields

preferredBackupWindow :: Maybe Text

The daily time range during which automated backups are created for your new database if automated backups are enabled.

The default is a 30-minute window selected at random from an 8-hour block of time for each AWS Region. For more information about the preferred backup window time blocks for each region, see the Working With Backups guide in the Amazon Relational Database Service (Amazon RDS) documentation.

Constraints:

Must be in the hh24:mi-hh24:mi format.

Example: 16:00-16:30

Specified in Coordinated Universal Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

masterUserPassword :: Maybe (Sensitive Text)

The password for the master user. The password can include any printable ASCII character except "/", """, or "@". It cannot contain spaces.

MySQL

Constraints: Must contain from 8 to 41 characters.

PostgreSQL

Constraints: Must contain from 8 to 128 characters.

publiclyAccessible :: Maybe Bool

Specifies the accessibility options for your new database. A value of true specifies a database that is available to resources outside of your Lightsail account. A value of false specifies a database that is available only to your Lightsail resources in the same region as your database.

availabilityZone :: Maybe Text

The Availability Zone in which to create your new database. Use the us-east-2a case-sensitive format.

You can get a list of Availability Zones by using the get regions operation. Be sure to add the include relational database Availability Zones parameter to your request.

preferredMaintenanceWindow :: Maybe Text

The weekly time range during which system maintenance can occur on your new database.

The default is a 30-minute window selected at random from an 8-hour block of time for each AWS Region, occurring on a random day of the week.

Constraints:

Must be in the ddd:hh24:mi-ddd:hh24:mi format.

Valid days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Must be at least 30 minutes.

Specified in Coordinated Universal Time (UTC).

Example: Tue:17:00-Tue:17:30

tags :: Maybe [Tag]

The tag keys and optional values to add to the resource during create.

Use the TagResource action to tag a resource after it's created.

relationalDatabaseName :: Text

The name to use for your new Lightsail database resource.

Constraints:

Must contain from 2 to 255 alphanumeric characters, or hyphens.

The first and last character must be a letter or number.

relationalDatabaseBlueprintId :: Text

The blueprint ID for your new database. A blueprint describes the major engine version of a database.

You can get a list of database blueprints IDs by using the get relational database blueprints operation.

relationalDatabaseBundleId :: Text

The bundle ID for your new database. A bundle describes the performance specifications for your database.

You can get a list of database bundle IDs by using the get relational database bundles operation.

masterDatabaseName :: Text

The meaning of this parameter differs according to the database engine you use.

MySQL

The name of the database to create when the Lightsail database resource is created. If this parameter isn't specified, no database is created in the database resource.

Constraints:

Must contain 1 to 64 letters or numbers.

Must begin with a letter. Subsequent characters can be letters, underscores, or digits (0- 9).

Can't be a word reserved by the specified database engine.

For more information about reserved words in MySQL, see the Keywords and Reserved Words articles for MySQL 5.6, MySQL 5.7, and MySQL 8.0.

PostgreSQL

The name of the database to create when the Lightsail database resource is created. If this parameter isn't specified, a database named postgres is created in the database resource.

Constraints:

Must contain 1 to 63 letters or numbers.

Must begin with a letter. Subsequent characters can be letters, underscores, or digits (0- 9).

Can't be a word reserved by the specified database engine.

For more information about reserved words in PostgreSQL, see the SQL Key Words articles for PostgreSQL 9.6, PostgreSQL 10, PostgreSQL 11, and PostgreSQL 12.

masterUsername :: Text

The name for the master user.

MySQL

Constraints:

Required for MySQL.

Must be 1 to 16 letters or numbers. Can contain underscores.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

For more information about reserved words in MySQL 5.6 or 5.7, see the Keywords and Reserved Words articles for MySQL 5.6, MySQL 5.7, or MySQL 8.0.

PostgreSQL

Constraints:

Required for PostgreSQL.

Must be 1 to 63 letters or numbers. Can contain underscores.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

For more information about reserved words in MySQL 5.6 or 5.7, see the Keywords and Reserved Words articles for PostgreSQL 9.6, PostgreSQL 10, PostgreSQL 11, and PostgreSQL 12.

newCreateRelationalDatabase

Arguments

:: Text
$sel:relationalDatabaseName:CreateRelationalDatabase'

-> Text
$sel:relationalDatabaseBlueprintId:CreateRelationalDatabase'

-> Text
$sel:relationalDatabaseBundleId:CreateRelationalDatabase'

-> Text
$sel:masterDatabaseName:CreateRelationalDatabase'

-> Text
$sel:masterUsername:CreateRelationalDatabase'

-> CreateRelationalDatabase

Create a value of CreateRelationalDatabase with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:preferredBackupWindow:CreateRelationalDatabase', createRelationalDatabase_preferredBackupWindow - The daily time range during which automated backups are created for your new database if automated backups are enabled.

The default is a 30-minute window selected at random from an 8-hour block of time for each AWS Region. For more information about the preferred backup window time blocks for each region, see the Working With Backups guide in the Amazon Relational Database Service (Amazon RDS) documentation.

Constraints:

Must be in the hh24:mi-hh24:mi format.

Example: 16:00-16:30

Specified in Coordinated Universal Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

$sel:masterUserPassword:CreateRelationalDatabase', createRelationalDatabase_masterUserPassword - The password for the master user. The password can include any printable ASCII character except "/", """, or "@". It cannot contain spaces.

MySQL

Constraints: Must contain from 8 to 41 characters.

PostgreSQL

Constraints: Must contain from 8 to 128 characters.

$sel:publiclyAccessible:CreateRelationalDatabase', createRelationalDatabase_publiclyAccessible - Specifies the accessibility options for your new database. A value of true specifies a database that is available to resources outside of your Lightsail account. A value of false specifies a database that is available only to your Lightsail resources in the same region as your database.

$sel:availabilityZone:CreateRelationalDatabase', createRelationalDatabase_availabilityZone - The Availability Zone in which to create your new database. Use the us-east-2a case-sensitive format.

You can get a list of Availability Zones by using the get regions operation. Be sure to add the include relational database Availability Zones parameter to your request.

$sel:preferredMaintenanceWindow:CreateRelationalDatabase', createRelationalDatabase_preferredMaintenanceWindow - The weekly time range during which system maintenance can occur on your new database.

The default is a 30-minute window selected at random from an 8-hour block of time for each AWS Region, occurring on a random day of the week.

Constraints:

Must be in the ddd:hh24:mi-ddd:hh24:mi format.

Valid days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Must be at least 30 minutes.

Specified in Coordinated Universal Time (UTC).

Example: Tue:17:00-Tue:17:30

$sel:tags:CreateRelationalDatabase', createRelationalDatabase_tags - The tag keys and optional values to add to the resource during create.

Use the TagResource action to tag a resource after it's created.

$sel:relationalDatabaseName:CreateRelationalDatabase', createRelationalDatabase_relationalDatabaseName - The name to use for your new Lightsail database resource.

Constraints:

Must contain from 2 to 255 alphanumeric characters, or hyphens.

The first and last character must be a letter or number.

$sel:relationalDatabaseBlueprintId:CreateRelationalDatabase', createRelationalDatabase_relationalDatabaseBlueprintId - The blueprint ID for your new database. A blueprint describes the major engine version of a database.

You can get a list of database blueprints IDs by using the get relational database blueprints operation.

$sel:relationalDatabaseBundleId:CreateRelationalDatabase', createRelationalDatabase_relationalDatabaseBundleId - The bundle ID for your new database. A bundle describes the performance specifications for your database.

You can get a list of database bundle IDs by using the get relational database bundles operation.

$sel:masterDatabaseName:CreateRelationalDatabase', createRelationalDatabase_masterDatabaseName - The meaning of this parameter differs according to the database engine you use.

MySQL

The name of the database to create when the Lightsail database resource is created. If this parameter isn't specified, no database is created in the database resource.

Constraints:

Must contain 1 to 64 letters or numbers.

Must begin with a letter. Subsequent characters can be letters, underscores, or digits (0- 9).

Can't be a word reserved by the specified database engine.

For more information about reserved words in MySQL, see the Keywords and Reserved Words articles for MySQL 5.6, MySQL 5.7, and MySQL 8.0.

PostgreSQL

The name of the database to create when the Lightsail database resource is created. If this parameter isn't specified, a database named postgres is created in the database resource.

Constraints:

Must contain 1 to 63 letters or numbers.

Must begin with a letter. Subsequent characters can be letters, underscores, or digits (0- 9).

Can't be a word reserved by the specified database engine.

For more information about reserved words in PostgreSQL, see the SQL Key Words articles for PostgreSQL 9.6, PostgreSQL 10, PostgreSQL 11, and PostgreSQL 12.

$sel:masterUsername:CreateRelationalDatabase', createRelationalDatabase_masterUsername - The name for the master user.

MySQL

Constraints:

Required for MySQL.

Must be 1 to 16 letters or numbers. Can contain underscores.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

For more information about reserved words in MySQL 5.6 or 5.7, see the Keywords and Reserved Words articles for MySQL 5.6, MySQL 5.7, or MySQL 8.0.

PostgreSQL

Constraints:

Required for PostgreSQL.

Must be 1 to 63 letters or numbers. Can contain underscores.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

For more information about reserved words in MySQL 5.6 or 5.7, see the Keywords and Reserved Words articles for PostgreSQL 9.6, PostgreSQL 10, PostgreSQL 11, and PostgreSQL 12.

Request Lenses

createRelationalDatabase_preferredBackupWindow :: Lens' CreateRelationalDatabase (Maybe Text)

The daily time range during which automated backups are created for your new database if automated backups are enabled.

The default is a 30-minute window selected at random from an 8-hour block of time for each AWS Region. For more information about the preferred backup window time blocks for each region, see the Working With Backups guide in the Amazon Relational Database Service (Amazon RDS) documentation.

Constraints:

Must be in the hh24:mi-hh24:mi format.

Example: 16:00-16:30

Specified in Coordinated Universal Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

createRelationalDatabase_masterUserPassword :: Lens' CreateRelationalDatabase (Maybe Text)

The password for the master user. The password can include any printable ASCII character except "/", """, or "@". It cannot contain spaces.

MySQL

Constraints: Must contain from 8 to 41 characters.

PostgreSQL

Constraints: Must contain from 8 to 128 characters.

createRelationalDatabase_publiclyAccessible :: Lens' CreateRelationalDatabase (Maybe Bool)

Specifies the accessibility options for your new database. A value of true specifies a database that is available to resources outside of your Lightsail account. A value of false specifies a database that is available only to your Lightsail resources in the same region as your database.

createRelationalDatabase_availabilityZone :: Lens' CreateRelationalDatabase (Maybe Text)

The Availability Zone in which to create your new database. Use the us-east-2a case-sensitive format.

You can get a list of Availability Zones by using the get regions operation. Be sure to add the include relational database Availability Zones parameter to your request.

createRelationalDatabase_preferredMaintenanceWindow :: Lens' CreateRelationalDatabase (Maybe Text)

The weekly time range during which system maintenance can occur on your new database.

The default is a 30-minute window selected at random from an 8-hour block of time for each AWS Region, occurring on a random day of the week.

Constraints:

Must be in the ddd:hh24:mi-ddd:hh24:mi format.

Valid days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Must be at least 30 minutes.

Specified in Coordinated Universal Time (UTC).

Example: Tue:17:00-Tue:17:30

createRelationalDatabase_tags :: Lens' CreateRelationalDatabase (Maybe [Tag])

The tag keys and optional values to add to the resource during create.

Use the TagResource action to tag a resource after it's created.

createRelationalDatabase_relationalDatabaseName :: Lens' CreateRelationalDatabase Text

The name to use for your new Lightsail database resource.

Constraints:

Must contain from 2 to 255 alphanumeric characters, or hyphens.

The first and last character must be a letter or number.

createRelationalDatabase_relationalDatabaseBlueprintId :: Lens' CreateRelationalDatabase Text

The blueprint ID for your new database. A blueprint describes the major engine version of a database.

You can get a list of database blueprints IDs by using the get relational database blueprints operation.

createRelationalDatabase_relationalDatabaseBundleId :: Lens' CreateRelationalDatabase Text

The bundle ID for your new database. A bundle describes the performance specifications for your database.

You can get a list of database bundle IDs by using the get relational database bundles operation.

createRelationalDatabase_masterDatabaseName :: Lens' CreateRelationalDatabase Text

The meaning of this parameter differs according to the database engine you use.

MySQL

The name of the database to create when the Lightsail database resource is created. If this parameter isn't specified, no database is created in the database resource.

Constraints:

Must contain 1 to 64 letters or numbers.

Must begin with a letter. Subsequent characters can be letters, underscores, or digits (0- 9).

Can't be a word reserved by the specified database engine.

For more information about reserved words in MySQL, see the Keywords and Reserved Words articles for MySQL 5.6, MySQL 5.7, and MySQL 8.0.

PostgreSQL

The name of the database to create when the Lightsail database resource is created. If this parameter isn't specified, a database named postgres is created in the database resource.

Constraints:

Must contain 1 to 63 letters or numbers.

Must begin with a letter. Subsequent characters can be letters, underscores, or digits (0- 9).

Can't be a word reserved by the specified database engine.

For more information about reserved words in PostgreSQL, see the SQL Key Words articles for PostgreSQL 9.6, PostgreSQL 10, PostgreSQL 11, and PostgreSQL 12.

createRelationalDatabase_masterUsername :: Lens' CreateRelationalDatabase Text

The name for the master user.

MySQL

Constraints:

Required for MySQL.

Must be 1 to 16 letters or numbers. Can contain underscores.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

For more information about reserved words in MySQL 5.6 or 5.7, see the Keywords and Reserved Words articles for MySQL 5.6, MySQL 5.7, or MySQL 8.0.

PostgreSQL

Constraints:

Required for PostgreSQL.

Must be 1 to 63 letters or numbers. Can contain underscores.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

For more information about reserved words in MySQL 5.6 or 5.7, see the Keywords and Reserved Words articles for PostgreSQL 9.6, PostgreSQL 10, PostgreSQL 11, and PostgreSQL 12.

Destructuring the Response

data CreateRelationalDatabaseResponse

See: newCreateRelationalDatabaseResponse smart constructor.

Constructors

CreateRelationalDatabaseResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newCreateRelationalDatabaseResponse

Arguments

:: Int
$sel:httpStatus:CreateRelationalDatabaseResponse'

-> CreateRelationalDatabaseResponse

Create a value of CreateRelationalDatabaseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:CreateRelationalDatabaseResponse', createRelationalDatabaseResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:CreateRelationalDatabaseResponse', createRelationalDatabaseResponse_httpStatus - The response's http status code.

Response Lenses

createRelationalDatabaseResponse_operations :: Lens' CreateRelationalDatabaseResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

createRelationalDatabaseResponse_httpStatus :: Lens' CreateRelationalDatabaseResponse Int

The response's http status code.