Network.AWS.Lightsail.DeleteAlarm

Description

Deletes an alarm.

An alarm is used to monitor a single metric for one of your resources. When a metric condition is met, the alarm can notify you by email, SMS text message, and a banner displayed on the Amazon Lightsail console. For more information, see Alarms in Amazon Lightsail.

Creating a Request

data DeleteAlarm

See: newDeleteAlarm smart constructor.

Constructors

DeleteAlarm'

Fields

alarmName :: Text

The name of the alarm to delete.

newDeleteAlarm

Arguments

:: Text
$sel:alarmName:DeleteAlarm'

-> DeleteAlarm

Create a value of DeleteAlarm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alarmName:DeleteAlarm', deleteAlarm_alarmName - The name of the alarm to delete.

Request Lenses

deleteAlarm_alarmName :: Lens' DeleteAlarm Text

The name of the alarm to delete.

Destructuring the Response

data DeleteAlarmResponse

See: newDeleteAlarmResponse smart constructor.

Constructors

DeleteAlarmResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newDeleteAlarmResponse

Arguments

:: Int
$sel:httpStatus:DeleteAlarmResponse'

-> DeleteAlarmResponse

Create a value of DeleteAlarmResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:DeleteAlarmResponse', deleteAlarmResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:DeleteAlarmResponse', deleteAlarmResponse_httpStatus - The response's http status code.

Response Lenses

deleteAlarmResponse_operations :: Lens' DeleteAlarmResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

deleteAlarmResponse_httpStatus :: Lens' DeleteAlarmResponse Int

The response's http status code.