Network.AWS.Lightsail.DeleteAutoSnapshot

Description

Deletes an automatic snapshot of an instance or disk. For more information, see the Amazon Lightsail Developer Guide.

Creating a Request

data DeleteAutoSnapshot

See: newDeleteAutoSnapshot smart constructor.

Constructors

DeleteAutoSnapshot'

Fields

resourceName :: Text

The name of the source instance or disk from which to delete the automatic snapshot.

date :: Text

The date of the automatic snapshot to delete in YYYY-MM-DD format. Use the get auto snapshots operation to get the available automatic snapshots for a resource.

newDeleteAutoSnapshot

Arguments

:: Text
$sel:resourceName:DeleteAutoSnapshot'

-> Text
$sel:date:DeleteAutoSnapshot'

-> DeleteAutoSnapshot

Create a value of DeleteAutoSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceName:DeleteAutoSnapshot', deleteAutoSnapshot_resourceName - The name of the source instance or disk from which to delete the automatic snapshot.

$sel:date:DeleteAutoSnapshot', deleteAutoSnapshot_date - The date of the automatic snapshot to delete in YYYY-MM-DD format. Use the get auto snapshots operation to get the available automatic snapshots for a resource.

Request Lenses

deleteAutoSnapshot_resourceName :: Lens' DeleteAutoSnapshot Text

The name of the source instance or disk from which to delete the automatic snapshot.

deleteAutoSnapshot_date :: Lens' DeleteAutoSnapshot Text

The date of the automatic snapshot to delete in YYYY-MM-DD format. Use the get auto snapshots operation to get the available automatic snapshots for a resource.

Destructuring the Response

data DeleteAutoSnapshotResponse

See: newDeleteAutoSnapshotResponse smart constructor.

Constructors

DeleteAutoSnapshotResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newDeleteAutoSnapshotResponse

Arguments

:: Int
$sel:httpStatus:DeleteAutoSnapshotResponse'

-> DeleteAutoSnapshotResponse

Create a value of DeleteAutoSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:DeleteAutoSnapshotResponse', deleteAutoSnapshotResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:DeleteAutoSnapshotResponse', deleteAutoSnapshotResponse_httpStatus - The response's http status code.

Response Lenses

deleteAutoSnapshotResponse_operations :: Lens' DeleteAutoSnapshotResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

deleteAutoSnapshotResponse_httpStatus :: Lens' DeleteAutoSnapshotResponse Int

The response's http status code.