Network.AWS.Lightsail.DeleteContactMethod

Description

Deletes a contact method.

A contact method is used to send you notifications about your Amazon Lightsail resources. You can add one email address and one mobile phone number contact method in each AWS Region. However, SMS text messaging is not supported in some AWS Regions, and SMS text messages cannot be sent to some countries/regions. For more information, see Notifications in Amazon Lightsail.

Creating a Request

data DeleteContactMethod

See: newDeleteContactMethod smart constructor.

Constructors

DeleteContactMethod'

Fields

protocol :: ContactProtocol

The protocol that will be deleted, such as Email or SMS (text messaging).

To delete an Email and an SMS contact method if you added both, you must run separate DeleteContactMethod actions to delete each protocol.

newDeleteContactMethod

Arguments

:: ContactProtocol
$sel:protocol:DeleteContactMethod'

-> DeleteContactMethod

Create a value of DeleteContactMethod with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protocol:DeleteContactMethod', deleteContactMethod_protocol - The protocol that will be deleted, such as Email or SMS (text messaging).

To delete an Email and an SMS contact method if you added both, you must run separate DeleteContactMethod actions to delete each protocol.

Request Lenses

deleteContactMethod_protocol :: Lens' DeleteContactMethod ContactProtocol

The protocol that will be deleted, such as Email or SMS (text messaging).

To delete an Email and an SMS contact method if you added both, you must run separate DeleteContactMethod actions to delete each protocol.

Destructuring the Response

data DeleteContactMethodResponse

See: newDeleteContactMethodResponse smart constructor.

Constructors

DeleteContactMethodResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newDeleteContactMethodResponse

Arguments

:: Int
$sel:httpStatus:DeleteContactMethodResponse'

-> DeleteContactMethodResponse

Create a value of DeleteContactMethodResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:DeleteContactMethodResponse', deleteContactMethodResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:DeleteContactMethodResponse', deleteContactMethodResponse_httpStatus - The response's http status code.

Response Lenses

deleteContactMethodResponse_operations :: Lens' DeleteContactMethodResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

deleteContactMethodResponse_httpStatus :: Lens' DeleteContactMethodResponse Int

The response's http status code.