Network.AWS.Lightsail.DeleteRelationalDatabase

Description

Deletes a database in Amazon Lightsail.

The delete relational database operation supports tag-based access control via resource tags applied to the resource identified by relationalDatabaseName. For more information, see the Amazon Lightsail Developer Guide.

Creating a Request

data DeleteRelationalDatabase

See: newDeleteRelationalDatabase smart constructor.

Constructors

DeleteRelationalDatabase'

Fields

skipFinalSnapshot :: Maybe Bool

Determines whether a final database snapshot is created before your database is deleted. If true is specified, no database snapshot is created. If false is specified, a database snapshot is created before your database is deleted.

You must specify the final relational database snapshot name parameter if the skip final snapshot parameter is false.

Default: false

finalRelationalDatabaseSnapshotName :: Maybe Text

The name of the database snapshot created if skip final snapshot is false, which is the default value for that parameter.

Specifying this parameter and also specifying the skip final snapshot parameter to true results in an error.

Constraints:

Must contain from 2 to 255 alphanumeric characters, or hyphens.

The first and last character must be a letter or number.

relationalDatabaseName :: Text

The name of the database that you are deleting.

newDeleteRelationalDatabase

Arguments

:: Text
$sel:relationalDatabaseName:DeleteRelationalDatabase'

-> DeleteRelationalDatabase

Create a value of DeleteRelationalDatabase with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:skipFinalSnapshot:DeleteRelationalDatabase', deleteRelationalDatabase_skipFinalSnapshot - Determines whether a final database snapshot is created before your database is deleted. If true is specified, no database snapshot is created. If false is specified, a database snapshot is created before your database is deleted.

You must specify the final relational database snapshot name parameter if the skip final snapshot parameter is false.

Default: false

$sel:finalRelationalDatabaseSnapshotName:DeleteRelationalDatabase', deleteRelationalDatabase_finalRelationalDatabaseSnapshotName - The name of the database snapshot created if skip final snapshot is false, which is the default value for that parameter.

Specifying this parameter and also specifying the skip final snapshot parameter to true results in an error.

Constraints:

Must contain from 2 to 255 alphanumeric characters, or hyphens.

The first and last character must be a letter or number.

$sel:relationalDatabaseName:DeleteRelationalDatabase', deleteRelationalDatabase_relationalDatabaseName - The name of the database that you are deleting.

Request Lenses

deleteRelationalDatabase_skipFinalSnapshot :: Lens' DeleteRelationalDatabase (Maybe Bool)

Determines whether a final database snapshot is created before your database is deleted. If true is specified, no database snapshot is created. If false is specified, a database snapshot is created before your database is deleted.

You must specify the final relational database snapshot name parameter if the skip final snapshot parameter is false.

Default: false

deleteRelationalDatabase_finalRelationalDatabaseSnapshotName :: Lens' DeleteRelationalDatabase (Maybe Text)

The name of the database snapshot created if skip final snapshot is false, which is the default value for that parameter.

Specifying this parameter and also specifying the skip final snapshot parameter to true results in an error.

Constraints:

Must contain from 2 to 255 alphanumeric characters, or hyphens.

The first and last character must be a letter or number.

deleteRelationalDatabase_relationalDatabaseName :: Lens' DeleteRelationalDatabase Text

The name of the database that you are deleting.

Destructuring the Response

data DeleteRelationalDatabaseResponse

See: newDeleteRelationalDatabaseResponse smart constructor.

Constructors

DeleteRelationalDatabaseResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newDeleteRelationalDatabaseResponse

Arguments

:: Int
$sel:httpStatus:DeleteRelationalDatabaseResponse'

-> DeleteRelationalDatabaseResponse

Create a value of DeleteRelationalDatabaseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:DeleteRelationalDatabaseResponse', deleteRelationalDatabaseResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:DeleteRelationalDatabaseResponse', deleteRelationalDatabaseResponse_httpStatus - The response's http status code.

Response Lenses

deleteRelationalDatabaseResponse_operations :: Lens' DeleteRelationalDatabaseResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

deleteRelationalDatabaseResponse_httpStatus :: Lens' DeleteRelationalDatabaseResponse Int

The response's http status code.