Network.AWS.Lightsail.DeleteRelationalDatabaseSnapshot

Description

Deletes a database snapshot in Amazon Lightsail.

The delete relational database snapshot operation supports tag-based access control via resource tags applied to the resource identified by relationalDatabaseName. For more information, see the Amazon Lightsail Developer Guide.

Creating a Request

data DeleteRelationalDatabaseSnapshot

See: newDeleteRelationalDatabaseSnapshot smart constructor.

Constructors

DeleteRelationalDatabaseSnapshot'

Fields

relationalDatabaseSnapshotName :: Text

The name of the database snapshot that you are deleting.

newDeleteRelationalDatabaseSnapshot

Arguments

:: Text
$sel:relationalDatabaseSnapshotName:DeleteRelationalDatabaseSnapshot'

-> DeleteRelationalDatabaseSnapshot

Create a value of DeleteRelationalDatabaseSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:relationalDatabaseSnapshotName:DeleteRelationalDatabaseSnapshot', deleteRelationalDatabaseSnapshot_relationalDatabaseSnapshotName - The name of the database snapshot that you are deleting.

Request Lenses

deleteRelationalDatabaseSnapshot_relationalDatabaseSnapshotName :: Lens' DeleteRelationalDatabaseSnapshot Text

The name of the database snapshot that you are deleting.

Destructuring the Response

data DeleteRelationalDatabaseSnapshotResponse

See: newDeleteRelationalDatabaseSnapshotResponse smart constructor.

Constructors

DeleteRelationalDatabaseSnapshotResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newDeleteRelationalDatabaseSnapshotResponse

Arguments

:: Int
$sel:httpStatus:DeleteRelationalDatabaseSnapshotResponse'

-> DeleteRelationalDatabaseSnapshotResponse

Create a value of DeleteRelationalDatabaseSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:DeleteRelationalDatabaseSnapshotResponse', deleteRelationalDatabaseSnapshotResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:DeleteRelationalDatabaseSnapshotResponse', deleteRelationalDatabaseSnapshotResponse_httpStatus - The response's http status code.

Response Lenses

deleteRelationalDatabaseSnapshotResponse_operations :: Lens' DeleteRelationalDatabaseSnapshotResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

deleteRelationalDatabaseSnapshotResponse_httpStatus :: Lens' DeleteRelationalDatabaseSnapshotResponse Int

The response's http status code.