Network.AWS.Lightsail.DetachStaticIp

Description

Detaches a static IP from the Amazon Lightsail instance to which it is attached.

Creating a Request

data DetachStaticIp

See: newDetachStaticIp smart constructor.

Constructors

DetachStaticIp'

Fields

staticIpName :: Text

The name of the static IP to detach from the instance.

newDetachStaticIp

Arguments

:: Text
$sel:staticIpName:DetachStaticIp'

-> DetachStaticIp

Create a value of DetachStaticIp with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:staticIpName:DetachStaticIp', detachStaticIp_staticIpName - The name of the static IP to detach from the instance.

Request Lenses

detachStaticIp_staticIpName :: Lens' DetachStaticIp Text

The name of the static IP to detach from the instance.

Destructuring the Response

data DetachStaticIpResponse

See: newDetachStaticIpResponse smart constructor.

Constructors

DetachStaticIpResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newDetachStaticIpResponse

Arguments

:: Int
$sel:httpStatus:DetachStaticIpResponse'

-> DetachStaticIpResponse

Create a value of DetachStaticIpResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:DetachStaticIpResponse', detachStaticIpResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:DetachStaticIpResponse', detachStaticIpResponse_httpStatus - The response's http status code.

Response Lenses

detachStaticIpResponse_operations :: Lens' DetachStaticIpResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

detachStaticIpResponse_httpStatus :: Lens' DetachStaticIpResponse Int

The response's http status code.