Network.AWS.Lightsail.DisableAddOn

Description

Disables an add-on for an Amazon Lightsail resource. For more information, see the Amazon Lightsail Developer Guide.

Creating a Request

data DisableAddOn

See: newDisableAddOn smart constructor.

Constructors

DisableAddOn'

Fields

addOnType :: AddOnType

The add-on type to disable.

resourceName :: Text

The name of the source resource for which to disable the add-on.

newDisableAddOn

Arguments

:: AddOnType
$sel:addOnType:DisableAddOn'

-> Text
$sel:resourceName:DisableAddOn'

-> DisableAddOn

Create a value of DisableAddOn with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:addOnType:DisableAddOn', disableAddOn_addOnType - The add-on type to disable.

$sel:resourceName:DisableAddOn', disableAddOn_resourceName - The name of the source resource for which to disable the add-on.

Request Lenses

disableAddOn_addOnType :: Lens' DisableAddOn AddOnType

The add-on type to disable.

disableAddOn_resourceName :: Lens' DisableAddOn Text

The name of the source resource for which to disable the add-on.

Destructuring the Response

data DisableAddOnResponse

See: newDisableAddOnResponse smart constructor.

Constructors

DisableAddOnResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newDisableAddOnResponse

Arguments

:: Int
$sel:httpStatus:DisableAddOnResponse'

-> DisableAddOnResponse

Create a value of DisableAddOnResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:DisableAddOnResponse', disableAddOnResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:DisableAddOnResponse', disableAddOnResponse_httpStatus - The response's http status code.

Response Lenses

disableAddOnResponse_operations :: Lens' DisableAddOnResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

disableAddOnResponse_httpStatus :: Lens' DisableAddOnResponse Int

The response's http status code.