Network.AWS.Lightsail.GetActiveNames

Description

Returns the names of all active (not deleted) resources.

This operation returns paginated results.

Creating a Request

data GetActiveNames

See: newGetActiveNames smart constructor.

Constructors

GetActiveNames'

Fields

pageToken :: Maybe Text

The token to advance to the next page of results from your request.

To get a page token, perform an initial GetActiveNames request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

newGetActiveNames :: GetActiveNames

Create a value of GetActiveNames with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pageToken:GetActiveNames', getActiveNames_pageToken - The token to advance to the next page of results from your request.

To get a page token, perform an initial GetActiveNames request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

Request Lenses

getActiveNames_pageToken :: Lens' GetActiveNames (Maybe Text)

The token to advance to the next page of results from your request.

To get a page token, perform an initial GetActiveNames request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

Destructuring the Response

data GetActiveNamesResponse

See: newGetActiveNamesResponse smart constructor.

Constructors

GetActiveNamesResponse'

Fields

nextPageToken :: Maybe Text

The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetActiveNames request and specify the next page token using the pageToken parameter.

activeNames :: Maybe [Text]

The list of active names returned by the get active names request.

httpStatus :: Int

The response's http status code.

newGetActiveNamesResponse

Arguments

:: Int
$sel:httpStatus:GetActiveNamesResponse'

-> GetActiveNamesResponse

Create a value of GetActiveNamesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:GetActiveNamesResponse', getActiveNamesResponse_nextPageToken - The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetActiveNames request and specify the next page token using the pageToken parameter.

$sel:activeNames:GetActiveNamesResponse', getActiveNamesResponse_activeNames - The list of active names returned by the get active names request.

$sel:httpStatus:GetActiveNamesResponse', getActiveNamesResponse_httpStatus - The response's http status code.

Response Lenses

getActiveNamesResponse_nextPageToken :: Lens' GetActiveNamesResponse (Maybe Text)

The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetActiveNames request and specify the next page token using the pageToken parameter.

getActiveNamesResponse_activeNames :: Lens' GetActiveNamesResponse (Maybe [Text])

The list of active names returned by the get active names request.

getActiveNamesResponse_httpStatus :: Lens' GetActiveNamesResponse Int

The response's http status code.