Network.AWS.Lightsail.GetAlarms

Description

Returns information about the configured alarms. Specify an alarm name in your request to return information about a specific alarm, or specify a monitored resource name to return information about all alarms for a specific resource.

An alarm is used to monitor a single metric for one of your resources. When a metric condition is met, the alarm can notify you by email, SMS text message, and a banner displayed on the Amazon Lightsail console. For more information, see Alarms in Amazon Lightsail.

Creating a Request

data GetAlarms

See: newGetAlarms smart constructor.

Constructors

GetAlarms'

Fields

pageToken :: Maybe Text

The token to advance to the next page of results from your request.

To get a page token, perform an initial GetAlarms request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

alarmName :: Maybe Text

The name of the alarm.

Specify an alarm name to return information about a specific alarm.

monitoredResourceName :: Maybe Text

The name of the Lightsail resource being monitored by the alarm.

Specify a monitored resource name to return information about all alarms for a specific resource.

newGetAlarms :: GetAlarms

Create a value of GetAlarms with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pageToken:GetAlarms', getAlarms_pageToken - The token to advance to the next page of results from your request.

To get a page token, perform an initial GetAlarms request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

$sel:alarmName:GetAlarms', getAlarms_alarmName - The name of the alarm.

Specify an alarm name to return information about a specific alarm.

$sel:monitoredResourceName:GetAlarms', getAlarms_monitoredResourceName - The name of the Lightsail resource being monitored by the alarm.

Specify a monitored resource name to return information about all alarms for a specific resource.

Request Lenses

getAlarms_pageToken :: Lens' GetAlarms (Maybe Text)

The token to advance to the next page of results from your request.

To get a page token, perform an initial GetAlarms request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

getAlarms_alarmName :: Lens' GetAlarms (Maybe Text)

The name of the alarm.

Specify an alarm name to return information about a specific alarm.

getAlarms_monitoredResourceName :: Lens' GetAlarms (Maybe Text)

The name of the Lightsail resource being monitored by the alarm.

Specify a monitored resource name to return information about all alarms for a specific resource.

Destructuring the Response

data GetAlarmsResponse

See: newGetAlarmsResponse smart constructor.

Constructors

GetAlarmsResponse'

Fields

nextPageToken :: Maybe Text

The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetAlarms request and specify the next page token using the pageToken parameter.

alarms :: Maybe [Alarm]

An array of objects that describe the alarms.

httpStatus :: Int

The response's http status code.

newGetAlarmsResponse

Arguments

:: Int
$sel:httpStatus:GetAlarmsResponse'

-> GetAlarmsResponse

Create a value of GetAlarmsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:GetAlarmsResponse', getAlarmsResponse_nextPageToken - The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetAlarms request and specify the next page token using the pageToken parameter.

$sel:alarms:GetAlarmsResponse', getAlarmsResponse_alarms - An array of objects that describe the alarms.

$sel:httpStatus:GetAlarmsResponse', getAlarmsResponse_httpStatus - The response's http status code.

Response Lenses

getAlarmsResponse_nextPageToken :: Lens' GetAlarmsResponse (Maybe Text)

The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetAlarms request and specify the next page token using the pageToken parameter.

getAlarmsResponse_alarms :: Lens' GetAlarmsResponse (Maybe [Alarm])

An array of objects that describe the alarms.

getAlarmsResponse_httpStatus :: Lens' GetAlarmsResponse Int

The response's http status code.