Network.AWS.Lightsail.GetBucketAccessKeys

Description

Returns the existing access key IDs for the specified Amazon Lightsail bucket.

This action does not return the secret access key value of an access key. You can get a secret access key only when you create it from the response of the CreateBucketAccessKey action. If you lose the secret access key, you must create a new access key.

Creating a Request

data GetBucketAccessKeys

See: newGetBucketAccessKeys smart constructor.

Constructors

GetBucketAccessKeys'

Fields

bucketName :: Text

The name of the bucket for which to return access keys.

newGetBucketAccessKeys

Arguments

:: Text
$sel:bucketName:GetBucketAccessKeys'

-> GetBucketAccessKeys

Create a value of GetBucketAccessKeys with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucketName:GetBucketAccessKeys', getBucketAccessKeys_bucketName - The name of the bucket for which to return access keys.

Request Lenses

getBucketAccessKeys_bucketName :: Lens' GetBucketAccessKeys Text

The name of the bucket for which to return access keys.

Destructuring the Response

data GetBucketAccessKeysResponse

See: newGetBucketAccessKeysResponse smart constructor.

Constructors

GetBucketAccessKeysResponse'

Fields

accessKeys :: Maybe [AccessKey]

An object that describes the access keys for the specified bucket.

httpStatus :: Int

The response's http status code.

newGetBucketAccessKeysResponse

Arguments

:: Int
$sel:httpStatus:GetBucketAccessKeysResponse'

-> GetBucketAccessKeysResponse

Create a value of GetBucketAccessKeysResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accessKeys:GetBucketAccessKeysResponse', getBucketAccessKeysResponse_accessKeys - An object that describes the access keys for the specified bucket.

$sel:httpStatus:GetBucketAccessKeysResponse', getBucketAccessKeysResponse_httpStatus - The response's http status code.

Response Lenses

getBucketAccessKeysResponse_accessKeys :: Lens' GetBucketAccessKeysResponse (Maybe [AccessKey])

An object that describes the access keys for the specified bucket.

getBucketAccessKeysResponse_httpStatus :: Lens' GetBucketAccessKeysResponse Int

The response's http status code.