Network.AWS.Lightsail.GetBundles

Description

Returns the list of bundles that are available for purchase. A bundle describes the specs for your virtual private server (or instance).

This operation returns paginated results.

Creating a Request

data GetBundles

See: newGetBundles smart constructor.

Constructors

GetBundles'

Fields

pageToken :: Maybe Text

The token to advance to the next page of results from your request.

To get a page token, perform an initial GetBundles request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

includeInactive :: Maybe Bool

A Boolean value that indicates whether to include inactive bundle results in your request.

newGetBundles :: GetBundles

Create a value of GetBundles with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pageToken:GetBundles', getBundles_pageToken - The token to advance to the next page of results from your request.

To get a page token, perform an initial GetBundles request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

$sel:includeInactive:GetBundles', getBundles_includeInactive - A Boolean value that indicates whether to include inactive bundle results in your request.

Request Lenses

getBundles_pageToken :: Lens' GetBundles (Maybe Text)

The token to advance to the next page of results from your request.

To get a page token, perform an initial GetBundles request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

getBundles_includeInactive :: Lens' GetBundles (Maybe Bool)

A Boolean value that indicates whether to include inactive bundle results in your request.

Destructuring the Response

data GetBundlesResponse

See: newGetBundlesResponse smart constructor.

Constructors

GetBundlesResponse'

Fields

nextPageToken :: Maybe Text

The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetBundles request and specify the next page token using the pageToken parameter.

bundles :: Maybe [Bundle]

An array of key-value pairs that contains information about the available bundles.

httpStatus :: Int

The response's http status code.

newGetBundlesResponse

Arguments

:: Int
$sel:httpStatus:GetBundlesResponse'

-> GetBundlesResponse

Create a value of GetBundlesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:GetBundlesResponse', getBundlesResponse_nextPageToken - The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetBundles request and specify the next page token using the pageToken parameter.

$sel:bundles:GetBundlesResponse', getBundlesResponse_bundles - An array of key-value pairs that contains information about the available bundles.

$sel:httpStatus:GetBundlesResponse', getBundlesResponse_httpStatus - The response's http status code.

Response Lenses

getBundlesResponse_nextPageToken :: Lens' GetBundlesResponse (Maybe Text)

The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetBundles request and specify the next page token using the pageToken parameter.

getBundlesResponse_bundles :: Lens' GetBundlesResponse (Maybe [Bundle])

An array of key-value pairs that contains information about the available bundles.

getBundlesResponse_httpStatus :: Lens' GetBundlesResponse Int

The response's http status code.