Network.AWS.Lightsail.GetCloudFormationStackRecords

Description

Returns the CloudFormation stack record created as a result of the create cloud formation stack operation.

An AWS CloudFormation stack is used to create a new Amazon EC2 instance from an exported Lightsail snapshot.

This operation returns paginated results.

Creating a Request

data GetCloudFormationStackRecords

See: newGetCloudFormationStackRecords smart constructor.

Constructors

GetCloudFormationStackRecords'

Fields

pageToken :: Maybe Text

The token to advance to the next page of results from your request.

To get a page token, perform an initial GetClouFormationStackRecords request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

newGetCloudFormationStackRecords :: GetCloudFormationStackRecords

Create a value of GetCloudFormationStackRecords with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pageToken:GetCloudFormationStackRecords', getCloudFormationStackRecords_pageToken - The token to advance to the next page of results from your request.

To get a page token, perform an initial GetClouFormationStackRecords request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

Request Lenses

getCloudFormationStackRecords_pageToken :: Lens' GetCloudFormationStackRecords (Maybe Text)

The token to advance to the next page of results from your request.

To get a page token, perform an initial GetClouFormationStackRecords request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

Destructuring the Response

data GetCloudFormationStackRecordsResponse

See: newGetCloudFormationStackRecordsResponse smart constructor.

Constructors

GetCloudFormationStackRecordsResponse'

Fields

cloudFormationStackRecords :: Maybe [CloudFormationStackRecord]

A list of objects describing the CloudFormation stack records.

nextPageToken :: Maybe Text

The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetCloudFormationStackRecords request and specify the next page token using the pageToken parameter.

httpStatus :: Int

The response's http status code.

newGetCloudFormationStackRecordsResponse

Arguments

:: Int
$sel:httpStatus:GetCloudFormationStackRecordsResponse'

-> GetCloudFormationStackRecordsResponse

Create a value of GetCloudFormationStackRecordsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cloudFormationStackRecords:GetCloudFormationStackRecordsResponse', getCloudFormationStackRecordsResponse_cloudFormationStackRecords - A list of objects describing the CloudFormation stack records.

$sel:nextPageToken:GetCloudFormationStackRecordsResponse', getCloudFormationStackRecordsResponse_nextPageToken - The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetCloudFormationStackRecords request and specify the next page token using the pageToken parameter.

$sel:httpStatus:GetCloudFormationStackRecordsResponse', getCloudFormationStackRecordsResponse_httpStatus - The response's http status code.

Response Lenses

getCloudFormationStackRecordsResponse_cloudFormationStackRecords :: Lens' GetCloudFormationStackRecordsResponse (Maybe [CloudFormationStackRecord])

A list of objects describing the CloudFormation stack records.

getCloudFormationStackRecordsResponse_nextPageToken :: Lens' GetCloudFormationStackRecordsResponse (Maybe Text)

The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetCloudFormationStackRecords request and specify the next page token using the pageToken parameter.

getCloudFormationStackRecordsResponse_httpStatus :: Lens' GetCloudFormationStackRecordsResponse Int

The response's http status code.