Network.AWS.Lightsail.GetContainerServiceMetricData

Description

Returns the data points of a specific metric of your Amazon Lightsail container service.

Metrics report the utilization of your resources. Monitor and collect metric data regularly to maintain the reliability, availability, and performance of your resources.

Creating a Request

data GetContainerServiceMetricData

See: newGetContainerServiceMetricData smart constructor.

Constructors

GetContainerServiceMetricData'

Fields

serviceName :: Text

The name of the container service for which to get metric data.

metricName :: ContainerServiceMetricName

The metric for which you want to return information.

Valid container service metric names are listed below, along with the most useful statistics to include in your request, and the published unit value.

CPUUtilization - The average percentage of compute units that are currently in use across all nodes of the container service. This metric identifies the processing power required to run containers on each node of the container service.

Statistics: The most useful statistics are Maximum and Average.

Unit: The published unit is Percent.

MemoryUtilization - The average percentage of available memory that is currently in use across all nodes of the container service. This metric identifies the memory required to run containers on each node of the container service.

Statistics: The most useful statistics are Maximum and Average.

Unit: The published unit is Percent.

startTime :: POSIX

The start time of the time period.

endTime :: POSIX

The end time of the time period.

period :: Natural

The granularity, in seconds, of the returned data points.

All container service metric data is available in 5-minute (300 seconds) granularity.

statistics :: [MetricStatistic]

The statistic for the metric.

The following statistics are available:

Minimum - The lowest value observed during the specified period. Use this value to determine low volumes of activity for your application.

Maximum - The highest value observed during the specified period. Use this value to determine high volumes of activity for your application.

Sum - All values submitted for the matching metric added together. You can use this statistic to determine the total volume of a metric.

Average - The value of Sum / SampleCount during the specified period. By comparing this statistic with the Minimum and Maximum values, you can determine the full scope of a metric and how close the average use is to the Minimum and Maximum values. This comparison helps you to know when to increase or decrease your resources.

SampleCount - The count, or number, of data points used for the statistical calculation.

newGetContainerServiceMetricData

Arguments

:: Text
$sel:serviceName:GetContainerServiceMetricData'

-> ContainerServiceMetricName
$sel:metricName:GetContainerServiceMetricData'

-> UTCTime
$sel:startTime:GetContainerServiceMetricData'

-> UTCTime
$sel:endTime:GetContainerServiceMetricData'

-> Natural
$sel:period:GetContainerServiceMetricData'

-> GetContainerServiceMetricData

Create a value of GetContainerServiceMetricData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceName:GetContainerServiceMetricData', getContainerServiceMetricData_serviceName - The name of the container service for which to get metric data.

$sel:metricName:GetContainerServiceMetricData', getContainerServiceMetricData_metricName - The metric for which you want to return information.

Valid container service metric names are listed below, along with the most useful statistics to include in your request, and the published unit value.

CPUUtilization - The average percentage of compute units that are currently in use across all nodes of the container service. This metric identifies the processing power required to run containers on each node of the container service.

Statistics: The most useful statistics are Maximum and Average.

Unit: The published unit is Percent.

MemoryUtilization - The average percentage of available memory that is currently in use across all nodes of the container service. This metric identifies the memory required to run containers on each node of the container service.

Statistics: The most useful statistics are Maximum and Average.

Unit: The published unit is Percent.

$sel:startTime:GetContainerServiceMetricData', getContainerServiceMetricData_startTime - The start time of the time period.

$sel:endTime:GetContainerServiceMetricData', getContainerServiceMetricData_endTime - The end time of the time period.

$sel:period:GetContainerServiceMetricData', getContainerServiceMetricData_period - The granularity, in seconds, of the returned data points.

All container service metric data is available in 5-minute (300 seconds) granularity.

$sel:statistics:GetContainerServiceMetricData', getContainerServiceMetricData_statistics - The statistic for the metric.

The following statistics are available:

Minimum - The lowest value observed during the specified period. Use this value to determine low volumes of activity for your application.

Maximum - The highest value observed during the specified period. Use this value to determine high volumes of activity for your application.

Sum - All values submitted for the matching metric added together. You can use this statistic to determine the total volume of a metric.

Average - The value of Sum / SampleCount during the specified period. By comparing this statistic with the Minimum and Maximum values, you can determine the full scope of a metric and how close the average use is to the Minimum and Maximum values. This comparison helps you to know when to increase or decrease your resources.

SampleCount - The count, or number, of data points used for the statistical calculation.

Request Lenses

getContainerServiceMetricData_serviceName :: Lens' GetContainerServiceMetricData Text

The name of the container service for which to get metric data.

getContainerServiceMetricData_metricName :: Lens' GetContainerServiceMetricData ContainerServiceMetricName

The metric for which you want to return information.

Valid container service metric names are listed below, along with the most useful statistics to include in your request, and the published unit value.

CPUUtilization - The average percentage of compute units that are currently in use across all nodes of the container service. This metric identifies the processing power required to run containers on each node of the container service.

Statistics: The most useful statistics are Maximum and Average.

Unit: The published unit is Percent.

MemoryUtilization - The average percentage of available memory that is currently in use across all nodes of the container service. This metric identifies the memory required to run containers on each node of the container service.

Statistics: The most useful statistics are Maximum and Average.

Unit: The published unit is Percent.

getContainerServiceMetricData_startTime :: Lens' GetContainerServiceMetricData UTCTime

The start time of the time period.

getContainerServiceMetricData_endTime :: Lens' GetContainerServiceMetricData UTCTime

The end time of the time period.

getContainerServiceMetricData_period :: Lens' GetContainerServiceMetricData Natural

The granularity, in seconds, of the returned data points.

All container service metric data is available in 5-minute (300 seconds) granularity.

getContainerServiceMetricData_statistics :: Lens' GetContainerServiceMetricData [MetricStatistic]

The statistic for the metric.

The following statistics are available:

Minimum - The lowest value observed during the specified period. Use this value to determine low volumes of activity for your application.

Maximum - The highest value observed during the specified period. Use this value to determine high volumes of activity for your application.

Sum - All values submitted for the matching metric added together. You can use this statistic to determine the total volume of a metric.

Average - The value of Sum / SampleCount during the specified period. By comparing this statistic with the Minimum and Maximum values, you can determine the full scope of a metric and how close the average use is to the Minimum and Maximum values. This comparison helps you to know when to increase or decrease your resources.

SampleCount - The count, or number, of data points used for the statistical calculation.

Destructuring the Response

data GetContainerServiceMetricDataResponse

See: newGetContainerServiceMetricDataResponse smart constructor.

Constructors

GetContainerServiceMetricDataResponse'

Fields

metricName :: Maybe ContainerServiceMetricName

The name of the metric returned.

metricData :: Maybe [MetricDatapoint]

An array of objects that describe the metric data returned.

httpStatus :: Int

The response's http status code.

newGetContainerServiceMetricDataResponse

Arguments

:: Int
$sel:httpStatus:GetContainerServiceMetricDataResponse'

-> GetContainerServiceMetricDataResponse

Create a value of GetContainerServiceMetricDataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricName:GetContainerServiceMetricData', getContainerServiceMetricDataResponse_metricName - The name of the metric returned.

$sel:metricData:GetContainerServiceMetricDataResponse', getContainerServiceMetricDataResponse_metricData - An array of objects that describe the metric data returned.

$sel:httpStatus:GetContainerServiceMetricDataResponse', getContainerServiceMetricDataResponse_httpStatus - The response's http status code.

Response Lenses

getContainerServiceMetricDataResponse_metricName :: Lens' GetContainerServiceMetricDataResponse (Maybe ContainerServiceMetricName)

The name of the metric returned.

getContainerServiceMetricDataResponse_metricData :: Lens' GetContainerServiceMetricDataResponse (Maybe [MetricDatapoint])

An array of objects that describe the metric data returned.

getContainerServiceMetricDataResponse_httpStatus :: Lens' GetContainerServiceMetricDataResponse Int

The response's http status code.