Network.AWS.Lightsail.GetDiskSnapshot

Description

Returns information about a specific block storage disk snapshot.

Creating a Request

data GetDiskSnapshot

See: newGetDiskSnapshot smart constructor.

Constructors

GetDiskSnapshot'

Fields

diskSnapshotName :: Text

The name of the disk snapshot (e.g., my-disk-snapshot).

newGetDiskSnapshot

Arguments

:: Text
$sel:diskSnapshotName:GetDiskSnapshot'

-> GetDiskSnapshot

Create a value of GetDiskSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:diskSnapshotName:GetDiskSnapshot', getDiskSnapshot_diskSnapshotName - The name of the disk snapshot (e.g., my-disk-snapshot).

Request Lenses

getDiskSnapshot_diskSnapshotName :: Lens' GetDiskSnapshot Text

The name of the disk snapshot (e.g., my-disk-snapshot).

Destructuring the Response

data GetDiskSnapshotResponse

See: newGetDiskSnapshotResponse smart constructor.

Constructors

GetDiskSnapshotResponse'

Fields

diskSnapshot :: Maybe DiskSnapshot

An object containing information about the disk snapshot.

httpStatus :: Int

The response's http status code.

newGetDiskSnapshotResponse

Arguments

:: Int
$sel:httpStatus:GetDiskSnapshotResponse'

-> GetDiskSnapshotResponse

Create a value of GetDiskSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:diskSnapshot:GetDiskSnapshotResponse', getDiskSnapshotResponse_diskSnapshot - An object containing information about the disk snapshot.

$sel:httpStatus:GetDiskSnapshotResponse', getDiskSnapshotResponse_httpStatus - The response's http status code.

Response Lenses

getDiskSnapshotResponse_diskSnapshot :: Lens' GetDiskSnapshotResponse (Maybe DiskSnapshot)

An object containing information about the disk snapshot.

getDiskSnapshotResponse_httpStatus :: Lens' GetDiskSnapshotResponse Int

The response's http status code.