Network.AWS.Lightsail.GetDistributionMetricData

Description

Returns the data points of a specific metric for an Amazon Lightsail content delivery network (CDN) distribution.

Metrics report the utilization of your resources, and the error counts generated by them. Monitor and collect metric data regularly to maintain the reliability, availability, and performance of your resources.

Creating a Request

data GetDistributionMetricData

See: newGetDistributionMetricData smart constructor.

Constructors

GetDistributionMetricData'

Fields

distributionName :: Text

The name of the distribution for which to get metric data.

Use the GetDistributions action to get a list of distribution names that you can specify.

metricName :: DistributionMetricName

The metric for which you want to return information.

Valid distribution metric names are listed below, along with the most useful statistics to include in your request, and the published unit value.

Requests - The total number of viewer requests received by your Lightsail distribution, for all HTTP methods, and for both HTTP and HTTPS requests.

Statistics: The most useful statistic is Sum.

Unit: The published unit is None.

BytesDownloaded - The number of bytes downloaded by viewers for GET, HEAD, and OPTIONS requests.

Statistics: The most useful statistic is Sum.

Unit: The published unit is None.

BytesUploaded - The number of bytes uploaded to your origin by your Lightsail distribution, using POST and PUT requests.

Statistics: The most useful statistic is Sum.

Unit: The published unit is None.

TotalErrorRate - The percentage of all viewer requests for which the response's HTTP status code was 4xx or 5xx.

Statistics: The most useful statistic is Average.

Unit: The published unit is Percent.

4xxErrorRate - The percentage of all viewer requests for which the response's HTTP status cod was 4xx. In these cases, the client or client viewer may have made an error. For example, a status code of 404 (Not Found) means that the client requested an object that could not be found.

Statistics: The most useful statistic is Average.

Unit: The published unit is Percent.

5xxErrorRate - The percentage of all viewer requests for which the response's HTTP status code was 5xx. In these cases, the origin server did not satisfy the requests. For example, a status code of 503 (Service Unavailable) means that the origin server is currently unavailable.

Statistics: The most useful statistic is Average.

Unit: The published unit is Percent.

startTime :: POSIX

The start of the time interval for which to get metric data.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use a start time of October 1, 2018, at 8 PM UTC, specify 1538424000 as the start time.

You can convert a human-friendly time to Unix time format using a converter like Epoch converter.

endTime :: POSIX

The end of the time interval for which to get metric data.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use an end time of October 1, 2018, at 9 PM UTC, specify 1538427600 as the end time.

You can convert a human-friendly time to Unix time format using a converter like Epoch converter.

period :: Natural

The granularity, in seconds, for the metric data points that will be returned.

unit :: MetricUnit

The unit for the metric data request.

Valid units depend on the metric data being requested. For the valid units with each available metric, see the metricName parameter.

statistics :: [MetricStatistic]

The statistic for the metric.

The following statistics are available:

Minimum - The lowest value observed during the specified period. Use this value to determine low volumes of activity for your application.

Maximum - The highest value observed during the specified period. Use this value to determine high volumes of activity for your application.

Sum - All values submitted for the matching metric added together. You can use this statistic to determine the total volume of a metric.

Average - The value of Sum / SampleCount during the specified period. By comparing this statistic with the Minimum and Maximum values, you can determine the full scope of a metric and how close the average use is to the Minimum and Maximum values. This comparison helps you to know when to increase or decrease your resources.

SampleCount - The count, or number, of data points used for the statistical calculation.

newGetDistributionMetricData

Arguments

:: Text
$sel:distributionName:GetDistributionMetricData'

-> DistributionMetricName
$sel:metricName:GetDistributionMetricData'

-> UTCTime
$sel:startTime:GetDistributionMetricData'

-> UTCTime
$sel:endTime:GetDistributionMetricData'

-> Natural
$sel:period:GetDistributionMetricData'

-> MetricUnit
$sel:unit:GetDistributionMetricData'

-> GetDistributionMetricData

Create a value of GetDistributionMetricData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:distributionName:GetDistributionMetricData', getDistributionMetricData_distributionName - The name of the distribution for which to get metric data.

Use the GetDistributions action to get a list of distribution names that you can specify.

$sel:metricName:GetDistributionMetricData', getDistributionMetricData_metricName - The metric for which you want to return information.

Valid distribution metric names are listed below, along with the most useful statistics to include in your request, and the published unit value.

Requests - The total number of viewer requests received by your Lightsail distribution, for all HTTP methods, and for both HTTP and HTTPS requests.

Statistics: The most useful statistic is Sum.

Unit: The published unit is None.

BytesDownloaded - The number of bytes downloaded by viewers for GET, HEAD, and OPTIONS requests.

Statistics: The most useful statistic is Sum.

Unit: The published unit is None.

BytesUploaded - The number of bytes uploaded to your origin by your Lightsail distribution, using POST and PUT requests.

Statistics: The most useful statistic is Sum.

Unit: The published unit is None.

TotalErrorRate - The percentage of all viewer requests for which the response's HTTP status code was 4xx or 5xx.

Statistics: The most useful statistic is Average.

Unit: The published unit is Percent.

4xxErrorRate - The percentage of all viewer requests for which the response's HTTP status cod was 4xx. In these cases, the client or client viewer may have made an error. For example, a status code of 404 (Not Found) means that the client requested an object that could not be found.

Statistics: The most useful statistic is Average.

Unit: The published unit is Percent.

5xxErrorRate - The percentage of all viewer requests for which the response's HTTP status code was 5xx. In these cases, the origin server did not satisfy the requests. For example, a status code of 503 (Service Unavailable) means that the origin server is currently unavailable.

Statistics: The most useful statistic is Average.

Unit: The published unit is Percent.

$sel:startTime:GetDistributionMetricData', getDistributionMetricData_startTime - The start of the time interval for which to get metric data.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use a start time of October 1, 2018, at 8 PM UTC, specify 1538424000 as the start time.

You can convert a human-friendly time to Unix time format using a converter like Epoch converter.

$sel:endTime:GetDistributionMetricData', getDistributionMetricData_endTime - The end of the time interval for which to get metric data.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use an end time of October 1, 2018, at 9 PM UTC, specify 1538427600 as the end time.

You can convert a human-friendly time to Unix time format using a converter like Epoch converter.

$sel:period:GetDistributionMetricData', getDistributionMetricData_period - The granularity, in seconds, for the metric data points that will be returned.

$sel:unit:GetDistributionMetricData', getDistributionMetricData_unit - The unit for the metric data request.

Valid units depend on the metric data being requested. For the valid units with each available metric, see the metricName parameter.

$sel:statistics:GetDistributionMetricData', getDistributionMetricData_statistics - The statistic for the metric.

The following statistics are available:

Minimum - The lowest value observed during the specified period. Use this value to determine low volumes of activity for your application.

Maximum - The highest value observed during the specified period. Use this value to determine high volumes of activity for your application.

Sum - All values submitted for the matching metric added together. You can use this statistic to determine the total volume of a metric.

Average - The value of Sum / SampleCount during the specified period. By comparing this statistic with the Minimum and Maximum values, you can determine the full scope of a metric and how close the average use is to the Minimum and Maximum values. This comparison helps you to know when to increase or decrease your resources.

SampleCount - The count, or number, of data points used for the statistical calculation.

Request Lenses

getDistributionMetricData_distributionName :: Lens' GetDistributionMetricData Text

The name of the distribution for which to get metric data.

Use the GetDistributions action to get a list of distribution names that you can specify.

getDistributionMetricData_metricName :: Lens' GetDistributionMetricData DistributionMetricName

The metric for which you want to return information.

Valid distribution metric names are listed below, along with the most useful statistics to include in your request, and the published unit value.

Requests - The total number of viewer requests received by your Lightsail distribution, for all HTTP methods, and for both HTTP and HTTPS requests.

Statistics: The most useful statistic is Sum.

Unit: The published unit is None.

BytesDownloaded - The number of bytes downloaded by viewers for GET, HEAD, and OPTIONS requests.

Statistics: The most useful statistic is Sum.

Unit: The published unit is None.

BytesUploaded - The number of bytes uploaded to your origin by your Lightsail distribution, using POST and PUT requests.

Statistics: The most useful statistic is Sum.

Unit: The published unit is None.

TotalErrorRate - The percentage of all viewer requests for which the response's HTTP status code was 4xx or 5xx.

Statistics: The most useful statistic is Average.

Unit: The published unit is Percent.

4xxErrorRate - The percentage of all viewer requests for which the response's HTTP status cod was 4xx. In these cases, the client or client viewer may have made an error. For example, a status code of 404 (Not Found) means that the client requested an object that could not be found.

Statistics: The most useful statistic is Average.

Unit: The published unit is Percent.

5xxErrorRate - The percentage of all viewer requests for which the response's HTTP status code was 5xx. In these cases, the origin server did not satisfy the requests. For example, a status code of 503 (Service Unavailable) means that the origin server is currently unavailable.

Statistics: The most useful statistic is Average.

Unit: The published unit is Percent.

getDistributionMetricData_startTime :: Lens' GetDistributionMetricData UTCTime

The start of the time interval for which to get metric data.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use a start time of October 1, 2018, at 8 PM UTC, specify 1538424000 as the start time.

You can convert a human-friendly time to Unix time format using a converter like Epoch converter.

getDistributionMetricData_endTime :: Lens' GetDistributionMetricData UTCTime

The end of the time interval for which to get metric data.

Constraints:

Specified in Coordinated Universal Time (UTC).

Specified in the Unix time format.

For example, if you wish to use an end time of October 1, 2018, at 9 PM UTC, specify 1538427600 as the end time.

You can convert a human-friendly time to Unix time format using a converter like Epoch converter.

getDistributionMetricData_period :: Lens' GetDistributionMetricData Natural

The granularity, in seconds, for the metric data points that will be returned.

getDistributionMetricData_unit :: Lens' GetDistributionMetricData MetricUnit

The unit for the metric data request.

Valid units depend on the metric data being requested. For the valid units with each available metric, see the metricName parameter.

getDistributionMetricData_statistics :: Lens' GetDistributionMetricData [MetricStatistic]

The statistic for the metric.

The following statistics are available:

Minimum - The lowest value observed during the specified period. Use this value to determine low volumes of activity for your application.

Maximum - The highest value observed during the specified period. Use this value to determine high volumes of activity for your application.

Sum - All values submitted for the matching metric added together. You can use this statistic to determine the total volume of a metric.

Average - The value of Sum / SampleCount during the specified period. By comparing this statistic with the Minimum and Maximum values, you can determine the full scope of a metric and how close the average use is to the Minimum and Maximum values. This comparison helps you to know when to increase or decrease your resources.

SampleCount - The count, or number, of data points used for the statistical calculation.

Destructuring the Response

data GetDistributionMetricDataResponse

See: newGetDistributionMetricDataResponse smart constructor.

Constructors

GetDistributionMetricDataResponse'

Fields

metricName :: Maybe DistributionMetricName

The name of the metric returned.

metricData :: Maybe [MetricDatapoint]

An array of objects that describe the metric data returned.

httpStatus :: Int

The response's http status code.

newGetDistributionMetricDataResponse

Arguments

:: Int
$sel:httpStatus:GetDistributionMetricDataResponse'

-> GetDistributionMetricDataResponse

Create a value of GetDistributionMetricDataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricName:GetDistributionMetricData', getDistributionMetricDataResponse_metricName - The name of the metric returned.

$sel:metricData:GetDistributionMetricDataResponse', getDistributionMetricDataResponse_metricData - An array of objects that describe the metric data returned.

$sel:httpStatus:GetDistributionMetricDataResponse', getDistributionMetricDataResponse_httpStatus - The response's http status code.

Response Lenses

getDistributionMetricDataResponse_metricName :: Lens' GetDistributionMetricDataResponse (Maybe DistributionMetricName)

The name of the metric returned.

getDistributionMetricDataResponse_metricData :: Lens' GetDistributionMetricDataResponse (Maybe [MetricDatapoint])

An array of objects that describe the metric data returned.

getDistributionMetricDataResponse_httpStatus :: Lens' GetDistributionMetricDataResponse Int

The response's http status code.