Network.AWS.Lightsail.GetInstanceMetricData

Description

Returns the data points for the specified Amazon Lightsail instance metric, given an instance name.

Metrics report the utilization of your resources, and the error counts generated by them. Monitor and collect metric data regularly to maintain the reliability, availability, and performance of your resources.

Creating a Request

data GetInstanceMetricData

See: newGetInstanceMetricData smart constructor.

Constructors

GetInstanceMetricData'

Fields

instanceName :: Text

The name of the instance for which you want to get metrics data.

metricName :: InstanceMetricName

The metric for which you want to return information.

Valid instance metric names are listed below, along with the most useful statistics to include in your request, and the published unit value.

BurstCapacityPercentage - The percentage of CPU performance available for your instance to burst above its baseline. Your instance continuously accrues and consumes burst capacity. Burst capacity stops accruing when your instance's BurstCapacityPercentage reaches 100%. For more information, see Viewing instance burst capacity in Amazon Lightsail.

Statistics: The most useful statistics are Maximum and Average.

Unit: The published unit is Percent.

BurstCapacityTime - The available amount of time for your instance to burst at 100% CPU utilization. Your instance continuously accrues and consumes burst capacity. Burst capacity time stops accruing when your instance's BurstCapacityPercentage metric reaches 100%.

Burst capacity time is consumed at the full rate only when your instance operates at 100% CPU utilization. For example, if your instance operates at 50% CPU utilization in the burstable zone for a 5-minute period, then it consumes CPU burst capacity minutes at a 50% rate in that period. Your instance consumed 2 minutes and 30 seconds of CPU burst capacity minutes in the 5-minute period. For more information, see Viewing instance burst capacity in Amazon Lightsail.

Statistics: The most useful statistics are Maximum and Average.

Unit: The published unit is Seconds.

CPUUtilization - The percentage of allocated compute units that are currently in use on the instance. This metric identifies the processing power to run the applications on the instance. Tools in your operating system can show a lower percentage than Lightsail when the instance is not allocated a full processor core.

Statistics: The most useful statistics are Maximum and Average.

Unit: The published unit is Percent.

NetworkIn - The number of bytes received on all network interfaces by the instance. This metric identifies the volume of incoming network traffic to the instance. The number reported is the number of bytes received during the period. Because this metric is reported in 5-minute intervals, divide the reported number by 300 to find Bytes/second.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Bytes.

NetworkOut - The number of bytes sent out on all network interfaces by the instance. This metric identifies the volume of outgoing network traffic from the instance. The number reported is the number of bytes sent during the period. Because this metric is reported in 5-minute intervals, divide the reported number by 300 to find Bytes/second.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Bytes.

StatusCheckFailed - Reports whether the instance passed or failed both the instance status check and the system status check. This metric can be either 0 (passed) or 1 (failed). This metric data is available in 1-minute (60 seconds) granularity.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Count.

StatusCheckFailed_Instance - Reports whether the instance passed or failed the instance status check. This metric can be either 0 (passed) or 1 (failed). This metric data is available in 1-minute (60 seconds) granularity.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Count.

StatusCheckFailed_System - Reports whether the instance passed or failed the system status check. This metric can be either 0 (passed) or 1 (failed). This metric data is available in 1-minute (60 seconds) granularity.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Count.

period :: Natural

The granularity, in seconds, of the returned data points.

The StatusCheckFailed, StatusCheckFailed_Instance, and StatusCheckFailed_System instance metric data is available in 1-minute (60 seconds) granularity. All other instance metric data is available in 5-minute (300 seconds) granularity.

startTime :: POSIX

The start time of the time period.

endTime :: POSIX

The end time of the time period.

unit :: MetricUnit

The unit for the metric data request. Valid units depend on the metric data being requested. For the valid units to specify with each available metric, see the metricName parameter.

statistics :: [MetricStatistic]

The statistic for the metric.

The following statistics are available:

Minimum - The lowest value observed during the specified period. Use this value to determine low volumes of activity for your application.

Maximum - The highest value observed during the specified period. Use this value to determine high volumes of activity for your application.

Sum - All values submitted for the matching metric added together. You can use this statistic to determine the total volume of a metric.

Average - The value of Sum / SampleCount during the specified period. By comparing this statistic with the Minimum and Maximum values, you can determine the full scope of a metric and how close the average use is to the Minimum and Maximum values. This comparison helps you to know when to increase or decrease your resources.

SampleCount - The count, or number, of data points used for the statistical calculation.

newGetInstanceMetricData

Arguments

:: Text
$sel:instanceName:GetInstanceMetricData'

-> InstanceMetricName
$sel:metricName:GetInstanceMetricData'

-> Natural
$sel:period:GetInstanceMetricData'

-> UTCTime
$sel:startTime:GetInstanceMetricData'

-> UTCTime
$sel:endTime:GetInstanceMetricData'

-> MetricUnit
$sel:unit:GetInstanceMetricData'

-> GetInstanceMetricData

Create a value of GetInstanceMetricData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceName:GetInstanceMetricData', getInstanceMetricData_instanceName - The name of the instance for which you want to get metrics data.

$sel:metricName:GetInstanceMetricData', getInstanceMetricData_metricName - The metric for which you want to return information.

Valid instance metric names are listed below, along with the most useful statistics to include in your request, and the published unit value.

BurstCapacityPercentage - The percentage of CPU performance available for your instance to burst above its baseline. Your instance continuously accrues and consumes burst capacity. Burst capacity stops accruing when your instance's BurstCapacityPercentage reaches 100%. For more information, see Viewing instance burst capacity in Amazon Lightsail.

Statistics: The most useful statistics are Maximum and Average.

Unit: The published unit is Percent.

BurstCapacityTime - The available amount of time for your instance to burst at 100% CPU utilization. Your instance continuously accrues and consumes burst capacity. Burst capacity time stops accruing when your instance's BurstCapacityPercentage metric reaches 100%.

Burst capacity time is consumed at the full rate only when your instance operates at 100% CPU utilization. For example, if your instance operates at 50% CPU utilization in the burstable zone for a 5-minute period, then it consumes CPU burst capacity minutes at a 50% rate in that period. Your instance consumed 2 minutes and 30 seconds of CPU burst capacity minutes in the 5-minute period. For more information, see Viewing instance burst capacity in Amazon Lightsail.

Statistics: The most useful statistics are Maximum and Average.

Unit: The published unit is Seconds.

CPUUtilization - The percentage of allocated compute units that are currently in use on the instance. This metric identifies the processing power to run the applications on the instance. Tools in your operating system can show a lower percentage than Lightsail when the instance is not allocated a full processor core.

Statistics: The most useful statistics are Maximum and Average.

Unit: The published unit is Percent.

NetworkIn - The number of bytes received on all network interfaces by the instance. This metric identifies the volume of incoming network traffic to the instance. The number reported is the number of bytes received during the period. Because this metric is reported in 5-minute intervals, divide the reported number by 300 to find Bytes/second.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Bytes.

NetworkOut - The number of bytes sent out on all network interfaces by the instance. This metric identifies the volume of outgoing network traffic from the instance. The number reported is the number of bytes sent during the period. Because this metric is reported in 5-minute intervals, divide the reported number by 300 to find Bytes/second.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Bytes.

StatusCheckFailed - Reports whether the instance passed or failed both the instance status check and the system status check. This metric can be either 0 (passed) or 1 (failed). This metric data is available in 1-minute (60 seconds) granularity.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Count.

StatusCheckFailed_Instance - Reports whether the instance passed or failed the instance status check. This metric can be either 0 (passed) or 1 (failed). This metric data is available in 1-minute (60 seconds) granularity.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Count.

StatusCheckFailed_System - Reports whether the instance passed or failed the system status check. This metric can be either 0 (passed) or 1 (failed). This metric data is available in 1-minute (60 seconds) granularity.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Count.

$sel:period:GetInstanceMetricData', getInstanceMetricData_period - The granularity, in seconds, of the returned data points.

The StatusCheckFailed, StatusCheckFailed_Instance, and StatusCheckFailed_System instance metric data is available in 1-minute (60 seconds) granularity. All other instance metric data is available in 5-minute (300 seconds) granularity.

$sel:startTime:GetInstanceMetricData', getInstanceMetricData_startTime - The start time of the time period.

$sel:endTime:GetInstanceMetricData', getInstanceMetricData_endTime - The end time of the time period.

$sel:unit:GetInstanceMetricData', getInstanceMetricData_unit - The unit for the metric data request. Valid units depend on the metric data being requested. For the valid units to specify with each available metric, see the metricName parameter.

$sel:statistics:GetInstanceMetricData', getInstanceMetricData_statistics - The statistic for the metric.

The following statistics are available:

Minimum - The lowest value observed during the specified period. Use this value to determine low volumes of activity for your application.

Maximum - The highest value observed during the specified period. Use this value to determine high volumes of activity for your application.

Sum - All values submitted for the matching metric added together. You can use this statistic to determine the total volume of a metric.

Average - The value of Sum / SampleCount during the specified period. By comparing this statistic with the Minimum and Maximum values, you can determine the full scope of a metric and how close the average use is to the Minimum and Maximum values. This comparison helps you to know when to increase or decrease your resources.

SampleCount - The count, or number, of data points used for the statistical calculation.

Request Lenses

getInstanceMetricData_instanceName :: Lens' GetInstanceMetricData Text

The name of the instance for which you want to get metrics data.

getInstanceMetricData_metricName :: Lens' GetInstanceMetricData InstanceMetricName

The metric for which you want to return information.

Valid instance metric names are listed below, along with the most useful statistics to include in your request, and the published unit value.

BurstCapacityPercentage - The percentage of CPU performance available for your instance to burst above its baseline. Your instance continuously accrues and consumes burst capacity. Burst capacity stops accruing when your instance's BurstCapacityPercentage reaches 100%. For more information, see Viewing instance burst capacity in Amazon Lightsail.

Statistics: The most useful statistics are Maximum and Average.

Unit: The published unit is Percent.

BurstCapacityTime - The available amount of time for your instance to burst at 100% CPU utilization. Your instance continuously accrues and consumes burst capacity. Burst capacity time stops accruing when your instance's BurstCapacityPercentage metric reaches 100%.

Burst capacity time is consumed at the full rate only when your instance operates at 100% CPU utilization. For example, if your instance operates at 50% CPU utilization in the burstable zone for a 5-minute period, then it consumes CPU burst capacity minutes at a 50% rate in that period. Your instance consumed 2 minutes and 30 seconds of CPU burst capacity minutes in the 5-minute period. For more information, see Viewing instance burst capacity in Amazon Lightsail.

Statistics: The most useful statistics are Maximum and Average.

Unit: The published unit is Seconds.

CPUUtilization - The percentage of allocated compute units that are currently in use on the instance. This metric identifies the processing power to run the applications on the instance. Tools in your operating system can show a lower percentage than Lightsail when the instance is not allocated a full processor core.

Statistics: The most useful statistics are Maximum and Average.

Unit: The published unit is Percent.

NetworkIn - The number of bytes received on all network interfaces by the instance. This metric identifies the volume of incoming network traffic to the instance. The number reported is the number of bytes received during the period. Because this metric is reported in 5-minute intervals, divide the reported number by 300 to find Bytes/second.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Bytes.

NetworkOut - The number of bytes sent out on all network interfaces by the instance. This metric identifies the volume of outgoing network traffic from the instance. The number reported is the number of bytes sent during the period. Because this metric is reported in 5-minute intervals, divide the reported number by 300 to find Bytes/second.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Bytes.

StatusCheckFailed - Reports whether the instance passed or failed both the instance status check and the system status check. This metric can be either 0 (passed) or 1 (failed). This metric data is available in 1-minute (60 seconds) granularity.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Count.

StatusCheckFailed_Instance - Reports whether the instance passed or failed the instance status check. This metric can be either 0 (passed) or 1 (failed). This metric data is available in 1-minute (60 seconds) granularity.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Count.

StatusCheckFailed_System - Reports whether the instance passed or failed the system status check. This metric can be either 0 (passed) or 1 (failed). This metric data is available in 1-minute (60 seconds) granularity.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Count.

getInstanceMetricData_period :: Lens' GetInstanceMetricData Natural

The granularity, in seconds, of the returned data points.

The StatusCheckFailed, StatusCheckFailed_Instance, and StatusCheckFailed_System instance metric data is available in 1-minute (60 seconds) granularity. All other instance metric data is available in 5-minute (300 seconds) granularity.

getInstanceMetricData_startTime :: Lens' GetInstanceMetricData UTCTime

The start time of the time period.

getInstanceMetricData_endTime :: Lens' GetInstanceMetricData UTCTime

The end time of the time period.

getInstanceMetricData_unit :: Lens' GetInstanceMetricData MetricUnit

The unit for the metric data request. Valid units depend on the metric data being requested. For the valid units to specify with each available metric, see the metricName parameter.

getInstanceMetricData_statistics :: Lens' GetInstanceMetricData [MetricStatistic]

The statistic for the metric.

The following statistics are available:

Minimum - The lowest value observed during the specified period. Use this value to determine low volumes of activity for your application.

Maximum - The highest value observed during the specified period. Use this value to determine high volumes of activity for your application.

Sum - All values submitted for the matching metric added together. You can use this statistic to determine the total volume of a metric.

Average - The value of Sum / SampleCount during the specified period. By comparing this statistic with the Minimum and Maximum values, you can determine the full scope of a metric and how close the average use is to the Minimum and Maximum values. This comparison helps you to know when to increase or decrease your resources.

SampleCount - The count, or number, of data points used for the statistical calculation.

Destructuring the Response

data GetInstanceMetricDataResponse

See: newGetInstanceMetricDataResponse smart constructor.

Constructors

GetInstanceMetricDataResponse'

Fields

metricName :: Maybe InstanceMetricName

The name of the metric returned.

metricData :: Maybe [MetricDatapoint]

An array of objects that describe the metric data returned.

httpStatus :: Int

The response's http status code.

newGetInstanceMetricDataResponse

Arguments

:: Int
$sel:httpStatus:GetInstanceMetricDataResponse'

-> GetInstanceMetricDataResponse

Create a value of GetInstanceMetricDataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricName:GetInstanceMetricData', getInstanceMetricDataResponse_metricName - The name of the metric returned.

$sel:metricData:GetInstanceMetricDataResponse', getInstanceMetricDataResponse_metricData - An array of objects that describe the metric data returned.

$sel:httpStatus:GetInstanceMetricDataResponse', getInstanceMetricDataResponse_httpStatus - The response's http status code.

Response Lenses

getInstanceMetricDataResponse_metricName :: Lens' GetInstanceMetricDataResponse (Maybe InstanceMetricName)

The name of the metric returned.

getInstanceMetricDataResponse_metricData :: Lens' GetInstanceMetricDataResponse (Maybe [MetricDatapoint])

An array of objects that describe the metric data returned.

getInstanceMetricDataResponse_httpStatus :: Lens' GetInstanceMetricDataResponse Int

The response's http status code.