Network.AWS.Lightsail.GetInstancePortStates

Description

Returns the firewall port states for a specific Amazon Lightsail instance, the IP addresses allowed to connect to the instance through the ports, and the protocol.

Creating a Request

data GetInstancePortStates

See: newGetInstancePortStates smart constructor.

Constructors

GetInstancePortStates'

Fields

instanceName :: Text

The name of the instance for which to return firewall port states.

newGetInstancePortStates

Arguments

:: Text
$sel:instanceName:GetInstancePortStates'

-> GetInstancePortStates

Create a value of GetInstancePortStates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceName:GetInstancePortStates', getInstancePortStates_instanceName - The name of the instance for which to return firewall port states.

Request Lenses

getInstancePortStates_instanceName :: Lens' GetInstancePortStates Text

The name of the instance for which to return firewall port states.

Destructuring the Response

data GetInstancePortStatesResponse

See: newGetInstancePortStatesResponse smart constructor.

Constructors

GetInstancePortStatesResponse'

Fields

portStates :: Maybe [InstancePortState]

An array of objects that describe the firewall port states for the specified instance.

httpStatus :: Int

The response's http status code.

newGetInstancePortStatesResponse

Arguments

:: Int
$sel:httpStatus:GetInstancePortStatesResponse'

-> GetInstancePortStatesResponse

Create a value of GetInstancePortStatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:portStates:GetInstancePortStatesResponse', getInstancePortStatesResponse_portStates - An array of objects that describe the firewall port states for the specified instance.

$sel:httpStatus:GetInstancePortStatesResponse', getInstancePortStatesResponse_httpStatus - The response's http status code.

Response Lenses

getInstancePortStatesResponse_portStates :: Lens' GetInstancePortStatesResponse (Maybe [InstancePortState])

An array of objects that describe the firewall port states for the specified instance.

getInstancePortStatesResponse_httpStatus :: Lens' GetInstancePortStatesResponse Int

The response's http status code.