Network.AWS.Lightsail.GetLoadBalancerMetricData

Description

Returns information about health metrics for your Lightsail load balancer.

Metrics report the utilization of your resources, and the error counts generated by them. Monitor and collect metric data regularly to maintain the reliability, availability, and performance of your resources.

Creating a Request

data GetLoadBalancerMetricData

See: newGetLoadBalancerMetricData smart constructor.

Constructors

GetLoadBalancerMetricData'

Fields

loadBalancerName :: Text

The name of the load balancer.

metricName :: LoadBalancerMetricName

The metric for which you want to return information.

Valid load balancer metric names are listed below, along with the most useful statistics to include in your request, and the published unit value.

ClientTLSNegotiationErrorCount - The number of TLS connections initiated by the client that did not establish a session with the load balancer due to a TLS error generated by the load balancer. Possible causes include a mismatch of ciphers or protocols.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Count.

HealthyHostCount - The number of target instances that are considered healthy.

Statistics: The most useful statistic are Average, Minimum, and Maximum.

Unit: The published unit is Count.

HTTPCode_Instance_2XX_Count - The number of HTTP 2XX response codes generated by the target instances. This does not include any response codes generated by the load balancer.

Statistics: The most useful statistic is Sum. Note that Minimum, Maximum, and Average all return 1.

Unit: The published unit is Count.

HTTPCode_Instance_3XX_Count - The number of HTTP 3XX response codes generated by the target instances. This does not include any response codes generated by the load balancer.

Statistics: The most useful statistic is Sum. Note that Minimum, Maximum, and Average all return 1.

Unit: The published unit is Count.

HTTPCode_Instance_4XX_Count - The number of HTTP 4XX response codes generated by the target instances. This does not include any response codes generated by the load balancer.

Statistics: The most useful statistic is Sum. Note that Minimum, Maximum, and Average all return 1.

Unit: The published unit is Count.

HTTPCode_Instance_5XX_Count - The number of HTTP 5XX response codes generated by the target instances. This does not include any response codes generated by the load balancer.

Statistics: The most useful statistic is Sum. Note that Minimum, Maximum, and Average all return 1.

Unit: The published unit is Count.

HTTPCode_LB_4XX_Count - The number of HTTP 4XX client error codes that originated from the load balancer. Client errors are generated when requests are malformed or incomplete. These requests were not received by the target instance. This count does not include response codes generated by the target instances.

Statistics: The most useful statistic is Sum. Note that Minimum, Maximum, and Average all return 1.

Unit: The published unit is Count.

HTTPCode_LB_5XX_Count - The number of HTTP 5XX server error codes that originated from the load balancer. This does not include any response codes generated by the target instance. This metric is reported if there are no healthy instances attached to the load balancer, or if the request rate exceeds the capacity of the instances (spillover) or the load balancer.

Statistics: The most useful statistic is Sum. Note that Minimum, Maximum, and Average all return 1.

Unit: The published unit is Count.

InstanceResponseTime - The time elapsed, in seconds, after the request leaves the load balancer until a response from the target instance is received.

Statistics: The most useful statistic is Average.

Unit: The published unit is Seconds.

RejectedConnectionCount - The number of connections that were rejected because the load balancer had reached its maximum number of connections.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Count.

RequestCount - The number of requests processed over IPv4. This count includes only the requests with a response generated by a target instance of the load balancer.

Statistics: The most useful statistic is Sum. Note that Minimum, Maximum, and Average all return 1.

Unit: The published unit is Count.

UnhealthyHostCount - The number of target instances that are considered unhealthy.

Statistics: The most useful statistic are Average, Minimum, and Maximum.

Unit: The published unit is Count.

period :: Natural

The granularity, in seconds, of the returned data points.

startTime :: POSIX

The start time of the period.

endTime :: POSIX

The end time of the period.

unit :: MetricUnit

The unit for the metric data request. Valid units depend on the metric data being requested. For the valid units with each available metric, see the metricName parameter.

statistics :: [MetricStatistic]

The statistic for the metric.

The following statistics are available:

Minimum - The lowest value observed during the specified period. Use this value to determine low volumes of activity for your application.

Maximum - The highest value observed during the specified period. Use this value to determine high volumes of activity for your application.

Sum - All values submitted for the matching metric added together. You can use this statistic to determine the total volume of a metric.

Average - The value of Sum / SampleCount during the specified period. By comparing this statistic with the Minimum and Maximum values, you can determine the full scope of a metric and how close the average use is to the Minimum and Maximum values. This comparison helps you to know when to increase or decrease your resources.

SampleCount - The count, or number, of data points used for the statistical calculation.

newGetLoadBalancerMetricData

Arguments

:: Text
$sel:loadBalancerName:GetLoadBalancerMetricData'

-> LoadBalancerMetricName
$sel:metricName:GetLoadBalancerMetricData'

-> Natural
$sel:period:GetLoadBalancerMetricData'

-> UTCTime
$sel:startTime:GetLoadBalancerMetricData'

-> UTCTime
$sel:endTime:GetLoadBalancerMetricData'

-> MetricUnit
$sel:unit:GetLoadBalancerMetricData'

-> GetLoadBalancerMetricData

Create a value of GetLoadBalancerMetricData with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loadBalancerName:GetLoadBalancerMetricData', getLoadBalancerMetricData_loadBalancerName - The name of the load balancer.

$sel:metricName:GetLoadBalancerMetricData', getLoadBalancerMetricData_metricName - The metric for which you want to return information.

Valid load balancer metric names are listed below, along with the most useful statistics to include in your request, and the published unit value.

ClientTLSNegotiationErrorCount - The number of TLS connections initiated by the client that did not establish a session with the load balancer due to a TLS error generated by the load balancer. Possible causes include a mismatch of ciphers or protocols.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Count.

HealthyHostCount - The number of target instances that are considered healthy.

Statistics: The most useful statistic are Average, Minimum, and Maximum.

Unit: The published unit is Count.

HTTPCode_Instance_2XX_Count - The number of HTTP 2XX response codes generated by the target instances. This does not include any response codes generated by the load balancer.

Statistics: The most useful statistic is Sum. Note that Minimum, Maximum, and Average all return 1.

Unit: The published unit is Count.

HTTPCode_Instance_3XX_Count - The number of HTTP 3XX response codes generated by the target instances. This does not include any response codes generated by the load balancer.

Statistics: The most useful statistic is Sum. Note that Minimum, Maximum, and Average all return 1.

Unit: The published unit is Count.

HTTPCode_Instance_4XX_Count - The number of HTTP 4XX response codes generated by the target instances. This does not include any response codes generated by the load balancer.

Statistics: The most useful statistic is Sum. Note that Minimum, Maximum, and Average all return 1.

Unit: The published unit is Count.

HTTPCode_Instance_5XX_Count - The number of HTTP 5XX response codes generated by the target instances. This does not include any response codes generated by the load balancer.

Statistics: The most useful statistic is Sum. Note that Minimum, Maximum, and Average all return 1.

Unit: The published unit is Count.

HTTPCode_LB_4XX_Count - The number of HTTP 4XX client error codes that originated from the load balancer. Client errors are generated when requests are malformed or incomplete. These requests were not received by the target instance. This count does not include response codes generated by the target instances.

Statistics: The most useful statistic is Sum. Note that Minimum, Maximum, and Average all return 1.

Unit: The published unit is Count.

HTTPCode_LB_5XX_Count - The number of HTTP 5XX server error codes that originated from the load balancer. This does not include any response codes generated by the target instance. This metric is reported if there are no healthy instances attached to the load balancer, or if the request rate exceeds the capacity of the instances (spillover) or the load balancer.

Statistics: The most useful statistic is Sum. Note that Minimum, Maximum, and Average all return 1.

Unit: The published unit is Count.

InstanceResponseTime - The time elapsed, in seconds, after the request leaves the load balancer until a response from the target instance is received.

Statistics: The most useful statistic is Average.

Unit: The published unit is Seconds.

RejectedConnectionCount - The number of connections that were rejected because the load balancer had reached its maximum number of connections.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Count.

RequestCount - The number of requests processed over IPv4. This count includes only the requests with a response generated by a target instance of the load balancer.

Statistics: The most useful statistic is Sum. Note that Minimum, Maximum, and Average all return 1.

Unit: The published unit is Count.

UnhealthyHostCount - The number of target instances that are considered unhealthy.

Statistics: The most useful statistic are Average, Minimum, and Maximum.

Unit: The published unit is Count.

$sel:period:GetLoadBalancerMetricData', getLoadBalancerMetricData_period - The granularity, in seconds, of the returned data points.

$sel:startTime:GetLoadBalancerMetricData', getLoadBalancerMetricData_startTime - The start time of the period.

$sel:endTime:GetLoadBalancerMetricData', getLoadBalancerMetricData_endTime - The end time of the period.

$sel:unit:GetLoadBalancerMetricData', getLoadBalancerMetricData_unit - The unit for the metric data request. Valid units depend on the metric data being requested. For the valid units with each available metric, see the metricName parameter.

$sel:statistics:GetLoadBalancerMetricData', getLoadBalancerMetricData_statistics - The statistic for the metric.

The following statistics are available:

Minimum - The lowest value observed during the specified period. Use this value to determine low volumes of activity for your application.

Maximum - The highest value observed during the specified period. Use this value to determine high volumes of activity for your application.

Sum - All values submitted for the matching metric added together. You can use this statistic to determine the total volume of a metric.

Average - The value of Sum / SampleCount during the specified period. By comparing this statistic with the Minimum and Maximum values, you can determine the full scope of a metric and how close the average use is to the Minimum and Maximum values. This comparison helps you to know when to increase or decrease your resources.

SampleCount - The count, or number, of data points used for the statistical calculation.

Request Lenses

getLoadBalancerMetricData_loadBalancerName :: Lens' GetLoadBalancerMetricData Text

The name of the load balancer.

getLoadBalancerMetricData_metricName :: Lens' GetLoadBalancerMetricData LoadBalancerMetricName

The metric for which you want to return information.

Valid load balancer metric names are listed below, along with the most useful statistics to include in your request, and the published unit value.

ClientTLSNegotiationErrorCount - The number of TLS connections initiated by the client that did not establish a session with the load balancer due to a TLS error generated by the load balancer. Possible causes include a mismatch of ciphers or protocols.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Count.

HealthyHostCount - The number of target instances that are considered healthy.

Statistics: The most useful statistic are Average, Minimum, and Maximum.

Unit: The published unit is Count.

HTTPCode_Instance_2XX_Count - The number of HTTP 2XX response codes generated by the target instances. This does not include any response codes generated by the load balancer.

Statistics: The most useful statistic is Sum. Note that Minimum, Maximum, and Average all return 1.

Unit: The published unit is Count.

HTTPCode_Instance_3XX_Count - The number of HTTP 3XX response codes generated by the target instances. This does not include any response codes generated by the load balancer.

Statistics: The most useful statistic is Sum. Note that Minimum, Maximum, and Average all return 1.

Unit: The published unit is Count.

HTTPCode_Instance_4XX_Count - The number of HTTP 4XX response codes generated by the target instances. This does not include any response codes generated by the load balancer.

Statistics: The most useful statistic is Sum. Note that Minimum, Maximum, and Average all return 1.

Unit: The published unit is Count.

HTTPCode_Instance_5XX_Count - The number of HTTP 5XX response codes generated by the target instances. This does not include any response codes generated by the load balancer.

Statistics: The most useful statistic is Sum. Note that Minimum, Maximum, and Average all return 1.

Unit: The published unit is Count.

HTTPCode_LB_4XX_Count - The number of HTTP 4XX client error codes that originated from the load balancer. Client errors are generated when requests are malformed or incomplete. These requests were not received by the target instance. This count does not include response codes generated by the target instances.

Statistics: The most useful statistic is Sum. Note that Minimum, Maximum, and Average all return 1.

Unit: The published unit is Count.

HTTPCode_LB_5XX_Count - The number of HTTP 5XX server error codes that originated from the load balancer. This does not include any response codes generated by the target instance. This metric is reported if there are no healthy instances attached to the load balancer, or if the request rate exceeds the capacity of the instances (spillover) or the load balancer.

Statistics: The most useful statistic is Sum. Note that Minimum, Maximum, and Average all return 1.

Unit: The published unit is Count.

InstanceResponseTime - The time elapsed, in seconds, after the request leaves the load balancer until a response from the target instance is received.

Statistics: The most useful statistic is Average.

Unit: The published unit is Seconds.

RejectedConnectionCount - The number of connections that were rejected because the load balancer had reached its maximum number of connections.

Statistics: The most useful statistic is Sum.

Unit: The published unit is Count.

RequestCount - The number of requests processed over IPv4. This count includes only the requests with a response generated by a target instance of the load balancer.

Statistics: The most useful statistic is Sum. Note that Minimum, Maximum, and Average all return 1.

Unit: The published unit is Count.

UnhealthyHostCount - The number of target instances that are considered unhealthy.

Statistics: The most useful statistic are Average, Minimum, and Maximum.

Unit: The published unit is Count.

getLoadBalancerMetricData_period :: Lens' GetLoadBalancerMetricData Natural

The granularity, in seconds, of the returned data points.

getLoadBalancerMetricData_startTime :: Lens' GetLoadBalancerMetricData UTCTime

The start time of the period.

getLoadBalancerMetricData_endTime :: Lens' GetLoadBalancerMetricData UTCTime

The end time of the period.

getLoadBalancerMetricData_unit :: Lens' GetLoadBalancerMetricData MetricUnit

The unit for the metric data request. Valid units depend on the metric data being requested. For the valid units with each available metric, see the metricName parameter.

getLoadBalancerMetricData_statistics :: Lens' GetLoadBalancerMetricData [MetricStatistic]

The statistic for the metric.

The following statistics are available:

Minimum - The lowest value observed during the specified period. Use this value to determine low volumes of activity for your application.

Maximum - The highest value observed during the specified period. Use this value to determine high volumes of activity for your application.

Sum - All values submitted for the matching metric added together. You can use this statistic to determine the total volume of a metric.

Average - The value of Sum / SampleCount during the specified period. By comparing this statistic with the Minimum and Maximum values, you can determine the full scope of a metric and how close the average use is to the Minimum and Maximum values. This comparison helps you to know when to increase or decrease your resources.

SampleCount - The count, or number, of data points used for the statistical calculation.

Destructuring the Response

data GetLoadBalancerMetricDataResponse

See: newGetLoadBalancerMetricDataResponse smart constructor.

Constructors

GetLoadBalancerMetricDataResponse'

Fields

metricName :: Maybe LoadBalancerMetricName

The name of the metric returned.

metricData :: Maybe [MetricDatapoint]

An array of objects that describe the metric data returned.

httpStatus :: Int

The response's http status code.

newGetLoadBalancerMetricDataResponse

Arguments

:: Int
$sel:httpStatus:GetLoadBalancerMetricDataResponse'

-> GetLoadBalancerMetricDataResponse

Create a value of GetLoadBalancerMetricDataResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricName:GetLoadBalancerMetricData', getLoadBalancerMetricDataResponse_metricName - The name of the metric returned.

$sel:metricData:GetLoadBalancerMetricDataResponse', getLoadBalancerMetricDataResponse_metricData - An array of objects that describe the metric data returned.

$sel:httpStatus:GetLoadBalancerMetricDataResponse', getLoadBalancerMetricDataResponse_httpStatus - The response's http status code.

Response Lenses

getLoadBalancerMetricDataResponse_metricName :: Lens' GetLoadBalancerMetricDataResponse (Maybe LoadBalancerMetricName)

The name of the metric returned.

getLoadBalancerMetricDataResponse_metricData :: Lens' GetLoadBalancerMetricDataResponse (Maybe [MetricDatapoint])

An array of objects that describe the metric data returned.

getLoadBalancerMetricDataResponse_httpStatus :: Lens' GetLoadBalancerMetricDataResponse Int

The response's http status code.