Network.AWS.Lightsail.GetOperationsForResource

Description

Gets operations for a specific resource (e.g., an instance or a static IP).

Creating a Request

data GetOperationsForResource

See: newGetOperationsForResource smart constructor.

Constructors

GetOperationsForResource'

Fields

pageToken :: Maybe Text

The token to advance to the next page of results from your request.

To get a page token, perform an initial GetOperationsForResource request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

resourceName :: Text

The name of the resource for which you are requesting information.

newGetOperationsForResource

Arguments

:: Text
$sel:resourceName:GetOperationsForResource'

-> GetOperationsForResource

Create a value of GetOperationsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pageToken:GetOperationsForResource', getOperationsForResource_pageToken - The token to advance to the next page of results from your request.

To get a page token, perform an initial GetOperationsForResource request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

$sel:resourceName:GetOperationsForResource', getOperationsForResource_resourceName - The name of the resource for which you are requesting information.

Request Lenses

getOperationsForResource_pageToken :: Lens' GetOperationsForResource (Maybe Text)

The token to advance to the next page of results from your request.

To get a page token, perform an initial GetOperationsForResource request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

getOperationsForResource_resourceName :: Lens' GetOperationsForResource Text

The name of the resource for which you are requesting information.

Destructuring the Response

data GetOperationsForResourceResponse

See: newGetOperationsForResourceResponse smart constructor.

Constructors

GetOperationsForResourceResponse'

Fields

nextPageCount :: Maybe Text

(Deprecated) Returns the number of pages of results that remain.

In releases prior to June 12, 2017, this parameter returned null by the API. It is now deprecated, and the API returns the next page token parameter instead.

nextPageToken :: Maybe Text

The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetOperationsForResource request and specify the next page token using the pageToken parameter.

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newGetOperationsForResourceResponse

Arguments

:: Int
$sel:httpStatus:GetOperationsForResourceResponse'

-> GetOperationsForResourceResponse

Create a value of GetOperationsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageCount:GetOperationsForResourceResponse', getOperationsForResourceResponse_nextPageCount - (Deprecated) Returns the number of pages of results that remain.

In releases prior to June 12, 2017, this parameter returned null by the API. It is now deprecated, and the API returns the next page token parameter instead.

$sel:nextPageToken:GetOperationsForResourceResponse', getOperationsForResourceResponse_nextPageToken - The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetOperationsForResource request and specify the next page token using the pageToken parameter.

$sel:operations:GetOperationsForResourceResponse', getOperationsForResourceResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:GetOperationsForResourceResponse', getOperationsForResourceResponse_httpStatus - The response's http status code.

Response Lenses

getOperationsForResourceResponse_nextPageCount :: Lens' GetOperationsForResourceResponse (Maybe Text)

(Deprecated) Returns the number of pages of results that remain.

In releases prior to June 12, 2017, this parameter returned null by the API. It is now deprecated, and the API returns the next page token parameter instead.

getOperationsForResourceResponse_nextPageToken :: Lens' GetOperationsForResourceResponse (Maybe Text)

The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetOperationsForResource request and specify the next page token using the pageToken parameter.

getOperationsForResourceResponse_operations :: Lens' GetOperationsForResourceResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

getOperationsForResourceResponse_httpStatus :: Lens' GetOperationsForResourceResponse Int

The response's http status code.