Network.AWS.Lightsail.GetRelationalDatabaseBlueprints

Description

Returns a list of available database blueprints in Amazon Lightsail. A blueprint describes the major engine version of a database.

You can use a blueprint ID to create a new database that runs a specific database engine.

This operation returns paginated results.

Creating a Request

data GetRelationalDatabaseBlueprints

See: newGetRelationalDatabaseBlueprints smart constructor.

Constructors

GetRelationalDatabaseBlueprints'

Fields

pageToken :: Maybe Text

The token to advance to the next page of results from your request.

To get a page token, perform an initial GetRelationalDatabaseBlueprints request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

newGetRelationalDatabaseBlueprints :: GetRelationalDatabaseBlueprints

Create a value of GetRelationalDatabaseBlueprints with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pageToken:GetRelationalDatabaseBlueprints', getRelationalDatabaseBlueprints_pageToken - The token to advance to the next page of results from your request.

To get a page token, perform an initial GetRelationalDatabaseBlueprints request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

Request Lenses

getRelationalDatabaseBlueprints_pageToken :: Lens' GetRelationalDatabaseBlueprints (Maybe Text)

The token to advance to the next page of results from your request.

To get a page token, perform an initial GetRelationalDatabaseBlueprints request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

Destructuring the Response

data GetRelationalDatabaseBlueprintsResponse

See: newGetRelationalDatabaseBlueprintsResponse smart constructor.

Constructors

GetRelationalDatabaseBlueprintsResponse'

Fields

blueprints :: Maybe [RelationalDatabaseBlueprint]

An object describing the result of your get relational database blueprints request.

nextPageToken :: Maybe Text

The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetRelationalDatabaseBlueprints request and specify the next page token using the pageToken parameter.

httpStatus :: Int

The response's http status code.

newGetRelationalDatabaseBlueprintsResponse

Arguments

:: Int
$sel:httpStatus:GetRelationalDatabaseBlueprintsResponse'

-> GetRelationalDatabaseBlueprintsResponse

Create a value of GetRelationalDatabaseBlueprintsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blueprints:GetRelationalDatabaseBlueprintsResponse', getRelationalDatabaseBlueprintsResponse_blueprints - An object describing the result of your get relational database blueprints request.

$sel:nextPageToken:GetRelationalDatabaseBlueprintsResponse', getRelationalDatabaseBlueprintsResponse_nextPageToken - The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetRelationalDatabaseBlueprints request and specify the next page token using the pageToken parameter.

$sel:httpStatus:GetRelationalDatabaseBlueprintsResponse', getRelationalDatabaseBlueprintsResponse_httpStatus - The response's http status code.

Response Lenses

getRelationalDatabaseBlueprintsResponse_blueprints :: Lens' GetRelationalDatabaseBlueprintsResponse (Maybe [RelationalDatabaseBlueprint])

An object describing the result of your get relational database blueprints request.

getRelationalDatabaseBlueprintsResponse_nextPageToken :: Lens' GetRelationalDatabaseBlueprintsResponse (Maybe Text)

The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetRelationalDatabaseBlueprints request and specify the next page token using the pageToken parameter.

getRelationalDatabaseBlueprintsResponse_httpStatus :: Lens' GetRelationalDatabaseBlueprintsResponse Int

The response's http status code.