Network.AWS.Lightsail.GetRelationalDatabaseParameters

Description

Returns all of the runtime parameters offered by the underlying database software, or engine, for a specific database in Amazon Lightsail.

In addition to the parameter names and values, this operation returns other information about each parameter. This information includes whether changes require a reboot, whether the parameter is modifiable, the allowed values, and the data types.

This operation returns paginated results.

Creating a Request

data GetRelationalDatabaseParameters

See: newGetRelationalDatabaseParameters smart constructor.

Constructors

GetRelationalDatabaseParameters'

Fields

pageToken :: Maybe Text

The token to advance to the next page of results from your request.

To get a page token, perform an initial GetRelationalDatabaseParameters request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

relationalDatabaseName :: Text

The name of your database for which to get parameters.

newGetRelationalDatabaseParameters

Arguments

:: Text
$sel:relationalDatabaseName:GetRelationalDatabaseParameters'

-> GetRelationalDatabaseParameters

Create a value of GetRelationalDatabaseParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pageToken:GetRelationalDatabaseParameters', getRelationalDatabaseParameters_pageToken - The token to advance to the next page of results from your request.

To get a page token, perform an initial GetRelationalDatabaseParameters request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

$sel:relationalDatabaseName:GetRelationalDatabaseParameters', getRelationalDatabaseParameters_relationalDatabaseName - The name of your database for which to get parameters.

Request Lenses

getRelationalDatabaseParameters_pageToken :: Lens' GetRelationalDatabaseParameters (Maybe Text)

The token to advance to the next page of results from your request.

To get a page token, perform an initial GetRelationalDatabaseParameters request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

getRelationalDatabaseParameters_relationalDatabaseName :: Lens' GetRelationalDatabaseParameters Text

The name of your database for which to get parameters.

Destructuring the Response

data GetRelationalDatabaseParametersResponse

See: newGetRelationalDatabaseParametersResponse smart constructor.

Constructors

GetRelationalDatabaseParametersResponse'

Fields

nextPageToken :: Maybe Text

The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetRelationalDatabaseParameters request and specify the next page token using the pageToken parameter.

parameters :: Maybe [RelationalDatabaseParameter]

An object describing the result of your get relational database parameters request.

httpStatus :: Int

The response's http status code.

newGetRelationalDatabaseParametersResponse

Arguments

:: Int
$sel:httpStatus:GetRelationalDatabaseParametersResponse'

-> GetRelationalDatabaseParametersResponse

Create a value of GetRelationalDatabaseParametersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextPageToken:GetRelationalDatabaseParametersResponse', getRelationalDatabaseParametersResponse_nextPageToken - The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetRelationalDatabaseParameters request and specify the next page token using the pageToken parameter.

$sel:parameters:GetRelationalDatabaseParametersResponse', getRelationalDatabaseParametersResponse_parameters - An object describing the result of your get relational database parameters request.

$sel:httpStatus:GetRelationalDatabaseParametersResponse', getRelationalDatabaseParametersResponse_httpStatus - The response's http status code.

Response Lenses

getRelationalDatabaseParametersResponse_nextPageToken :: Lens' GetRelationalDatabaseParametersResponse (Maybe Text)

The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetRelationalDatabaseParameters request and specify the next page token using the pageToken parameter.

getRelationalDatabaseParametersResponse_parameters :: Lens' GetRelationalDatabaseParametersResponse (Maybe [RelationalDatabaseParameter])

An object describing the result of your get relational database parameters request.

getRelationalDatabaseParametersResponse_httpStatus :: Lens' GetRelationalDatabaseParametersResponse Int

The response's http status code.