Network.AWS.Lightsail.GetRelationalDatabaseSnapshots

Description

Returns information about all of your database snapshots in Amazon Lightsail.

This operation returns paginated results.

Creating a Request

data GetRelationalDatabaseSnapshots

See: newGetRelationalDatabaseSnapshots smart constructor.

Constructors

GetRelationalDatabaseSnapshots'

Fields

pageToken :: Maybe Text

The token to advance to the next page of results from your request.

To get a page token, perform an initial GetRelationalDatabaseSnapshots request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

newGetRelationalDatabaseSnapshots :: GetRelationalDatabaseSnapshots

Create a value of GetRelationalDatabaseSnapshots with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pageToken:GetRelationalDatabaseSnapshots', getRelationalDatabaseSnapshots_pageToken - The token to advance to the next page of results from your request.

To get a page token, perform an initial GetRelationalDatabaseSnapshots request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

Request Lenses

getRelationalDatabaseSnapshots_pageToken :: Lens' GetRelationalDatabaseSnapshots (Maybe Text)

The token to advance to the next page of results from your request.

To get a page token, perform an initial GetRelationalDatabaseSnapshots request. If your results are paginated, the response will return a next page token that you can specify as the page token in a subsequent request.

Destructuring the Response

data GetRelationalDatabaseSnapshotsResponse

See: newGetRelationalDatabaseSnapshotsResponse smart constructor.

Constructors

GetRelationalDatabaseSnapshotsResponse'

Fields

relationalDatabaseSnapshots :: Maybe [RelationalDatabaseSnapshot]

An object describing the result of your get relational database snapshots request.

nextPageToken :: Maybe Text

The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetRelationalDatabaseSnapshots request and specify the next page token using the pageToken parameter.

httpStatus :: Int

The response's http status code.

newGetRelationalDatabaseSnapshotsResponse

Arguments

:: Int
$sel:httpStatus:GetRelationalDatabaseSnapshotsResponse'

-> GetRelationalDatabaseSnapshotsResponse

Create a value of GetRelationalDatabaseSnapshotsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:relationalDatabaseSnapshots:GetRelationalDatabaseSnapshotsResponse', getRelationalDatabaseSnapshotsResponse_relationalDatabaseSnapshots - An object describing the result of your get relational database snapshots request.

$sel:nextPageToken:GetRelationalDatabaseSnapshotsResponse', getRelationalDatabaseSnapshotsResponse_nextPageToken - The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetRelationalDatabaseSnapshots request and specify the next page token using the pageToken parameter.

$sel:httpStatus:GetRelationalDatabaseSnapshotsResponse', getRelationalDatabaseSnapshotsResponse_httpStatus - The response's http status code.

Response Lenses

getRelationalDatabaseSnapshotsResponse_relationalDatabaseSnapshots :: Lens' GetRelationalDatabaseSnapshotsResponse (Maybe [RelationalDatabaseSnapshot])

An object describing the result of your get relational database snapshots request.

getRelationalDatabaseSnapshotsResponse_nextPageToken :: Lens' GetRelationalDatabaseSnapshotsResponse (Maybe Text)

The token to advance to the next page of results from your request.

A next page token is not returned if there are no more results to display.

To get the next page of results, perform another GetRelationalDatabaseSnapshots request and specify the next page token using the pageToken parameter.

getRelationalDatabaseSnapshotsResponse_httpStatus :: Lens' GetRelationalDatabaseSnapshotsResponse Int

The response's http status code.