Network.AWS.Lightsail.OpenInstancePublicPorts

Description

Opens ports for a specific Amazon Lightsail instance, and specifies the IP addresses allowed to connect to the instance through the ports, and the protocol.

The OpenInstancePublicPorts action supports tag-based access control via resource tags applied to the resource identified by instanceName. For more information, see the Amazon Lightsail Developer Guide.

Creating a Request

data OpenInstancePublicPorts

See: newOpenInstancePublicPorts smart constructor.

Constructors

OpenInstancePublicPorts'

Fields

portInfo :: PortInfo

An object to describe the ports to open for the specified instance.

instanceName :: Text

The name of the instance for which to open ports.

newOpenInstancePublicPorts

Arguments

:: PortInfo
$sel:portInfo:OpenInstancePublicPorts'

-> Text
$sel:instanceName:OpenInstancePublicPorts'

-> OpenInstancePublicPorts

Create a value of OpenInstancePublicPorts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:portInfo:OpenInstancePublicPorts', openInstancePublicPorts_portInfo - An object to describe the ports to open for the specified instance.

$sel:instanceName:OpenInstancePublicPorts', openInstancePublicPorts_instanceName - The name of the instance for which to open ports.

Request Lenses

openInstancePublicPorts_portInfo :: Lens' OpenInstancePublicPorts PortInfo

An object to describe the ports to open for the specified instance.

openInstancePublicPorts_instanceName :: Lens' OpenInstancePublicPorts Text

The name of the instance for which to open ports.

Destructuring the Response

data OpenInstancePublicPortsResponse

See: newOpenInstancePublicPortsResponse smart constructor.

Constructors

OpenInstancePublicPortsResponse'

Fields

operation :: Maybe Operation

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newOpenInstancePublicPortsResponse

Arguments

:: Int
$sel:httpStatus:OpenInstancePublicPortsResponse'

-> OpenInstancePublicPortsResponse

Create a value of OpenInstancePublicPortsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operation:OpenInstancePublicPortsResponse', openInstancePublicPortsResponse_operation - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:OpenInstancePublicPortsResponse', openInstancePublicPortsResponse_httpStatus - The response's http status code.

Response Lenses

openInstancePublicPortsResponse_operation :: Lens' OpenInstancePublicPortsResponse (Maybe Operation)

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

openInstancePublicPortsResponse_httpStatus :: Lens' OpenInstancePublicPortsResponse Int

The response's http status code.