Network.AWS.Lightsail.PutAlarm

Description

Creates or updates an alarm, and associates it with the specified metric.

An alarm is used to monitor a single metric for one of your resources. When a metric condition is met, the alarm can notify you by email, SMS text message, and a banner displayed on the Amazon Lightsail console. For more information, see Alarms in Amazon Lightsail.

When this action creates an alarm, the alarm state is immediately set to INSUFFICIENT_DATA. The alarm is then evaluated and its state is set appropriately. Any actions associated with the new state are then executed.

When you update an existing alarm, its state is left unchanged, but the update completely overwrites the previous configuration of the alarm. The alarm is then evaluated with the updated configuration.

Creating a Request

data PutAlarm

See: newPutAlarm smart constructor.

Constructors

PutAlarm'

Fields

treatMissingData :: Maybe TreatMissingData

Sets how this alarm will handle missing data points.

An alarm can treat missing data in the following ways:

breaching - Assume the missing data is not within the threshold. Missing data counts towards the number of times the metric is not within the threshold.

notBreaching - Assume the missing data is within the threshold. Missing data does not count towards the number of times the metric is not within the threshold.

ignore - Ignore the missing data. Maintains the current alarm state.

missing - Missing data is treated as missing.

If treatMissingData is not specified, the default behavior of missing is used.

contactProtocols :: Maybe [ContactProtocol]

The contact protocols to use for the alarm, such as Email, SMS (text messaging), or both.

A notification is sent via the specified contact protocol if notifications are enabled for the alarm, and when the alarm is triggered.

A notification is not sent if a contact protocol is not specified, if the specified contact protocol is not configured in the AWS Region, or if notifications are not enabled for the alarm using the notificationEnabled paramater.

Use the CreateContactMethod action to configure a contact protocol in an AWS Region.

datapointsToAlarm :: Maybe Int

The number of data points that must be not within the specified threshold to trigger the alarm. If you are setting an "M out of N" alarm, this value (datapointsToAlarm) is the M.

notificationEnabled :: Maybe Bool

Indicates whether the alarm is enabled.

Notifications are enabled by default if you don't specify this parameter.

notificationTriggers :: Maybe [AlarmState]

The alarm states that trigger a notification.

An alarm has the following possible states:

ALARM - The metric is outside of the defined threshold.

INSUFFICIENT_DATA - The alarm has just started, the metric is not available, or not enough data is available for the metric to determine the alarm state.

OK - The metric is within the defined threshold.

When you specify a notification trigger, the ALARM state must be specified. The INSUFFICIENT_DATA and OK states can be specified in addition to the ALARM state.

If you specify OK as an alarm trigger, a notification is sent when the alarm switches from an ALARM or INSUFFICIENT_DATA alarm state to an OK state. This can be thought of as an all clear alarm notification.

If you specify INSUFFICIENT_DATA as the alarm trigger, a notification is sent when the alarm switches from an OK or ALARM alarm state to an INSUFFICIENT_DATA state.

The notification trigger defaults to ALARM if you don't specify this parameter.

alarmName :: Text

The name for the alarm. Specify the name of an existing alarm to update, and overwrite the previous configuration of the alarm.

metricName :: MetricName

The name of the metric to associate with the alarm.

You can configure up to two alarms per metric.

The following metrics are available for each resource type:

Instances: BurstCapacityPercentage, BurstCapacityTime, CPUUtilization, NetworkIn, NetworkOut, StatusCheckFailed, StatusCheckFailed_Instance, and StatusCheckFailed_System.

Load balancers: ClientTLSNegotiationErrorCount, HealthyHostCount, UnhealthyHostCount, HTTPCode_LB_4XX_Count, HTTPCode_LB_5XX_Count, HTTPCode_Instance_2XX_Count, HTTPCode_Instance_3XX_Count, HTTPCode_Instance_4XX_Count, HTTPCode_Instance_5XX_Count, InstanceResponseTime, RejectedConnectionCount, and RequestCount.

Relational databases: CPUUtilization, DatabaseConnections, DiskQueueDepth, FreeStorageSpace, NetworkReceiveThroughput, and NetworkTransmitThroughput.

For more information about these metrics, see Metrics available in Lightsail.

monitoredResourceName :: Text

The name of the Lightsail resource that will be monitored.

Instances, load balancers, and relational databases are the only Lightsail resources that can currently be monitored by alarms.

comparisonOperator :: ComparisonOperator

The arithmetic operation to use when comparing the specified statistic to the threshold. The specified statistic value is used as the first operand.

threshold :: Double

The value against which the specified statistic is compared.

evaluationPeriods :: Int

The number of most recent periods over which data is compared to the specified threshold. If you are setting an "M out of N" alarm, this value (evaluationPeriods) is the N.

If you are setting an alarm that requires that a number of consecutive data points be breaching to trigger the alarm, this value specifies the rolling period of time in which data points are evaluated.

Each evaluation period is five minutes long. For example, specify an evaluation period of 24 to evaluate a metric over a rolling period of two hours.

You can specify a minimum valuation period of 1 (5 minutes), and a maximum evaluation period of 288 (24 hours).

newPutAlarm

Arguments

:: Text
$sel:alarmName:PutAlarm'

-> MetricName
$sel:metricName:PutAlarm'

-> Text
$sel:monitoredResourceName:PutAlarm'

-> ComparisonOperator
$sel:comparisonOperator:PutAlarm'

-> Double
$sel:threshold:PutAlarm'

-> Int
$sel:evaluationPeriods:PutAlarm'

-> PutAlarm

Create a value of PutAlarm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:treatMissingData:PutAlarm', putAlarm_treatMissingData - Sets how this alarm will handle missing data points.

An alarm can treat missing data in the following ways:

breaching - Assume the missing data is not within the threshold. Missing data counts towards the number of times the metric is not within the threshold.

notBreaching - Assume the missing data is within the threshold. Missing data does not count towards the number of times the metric is not within the threshold.

ignore - Ignore the missing data. Maintains the current alarm state.

missing - Missing data is treated as missing.

If treatMissingData is not specified, the default behavior of missing is used.

$sel:contactProtocols:PutAlarm', putAlarm_contactProtocols - The contact protocols to use for the alarm, such as Email, SMS (text messaging), or both.

A notification is sent via the specified contact protocol if notifications are enabled for the alarm, and when the alarm is triggered.

A notification is not sent if a contact protocol is not specified, if the specified contact protocol is not configured in the AWS Region, or if notifications are not enabled for the alarm using the notificationEnabled paramater.

Use the CreateContactMethod action to configure a contact protocol in an AWS Region.

$sel:datapointsToAlarm:PutAlarm', putAlarm_datapointsToAlarm - The number of data points that must be not within the specified threshold to trigger the alarm. If you are setting an "M out of N" alarm, this value (datapointsToAlarm) is the M.

$sel:notificationEnabled:PutAlarm', putAlarm_notificationEnabled - Indicates whether the alarm is enabled.

Notifications are enabled by default if you don't specify this parameter.

$sel:notificationTriggers:PutAlarm', putAlarm_notificationTriggers - The alarm states that trigger a notification.

An alarm has the following possible states:

ALARM - The metric is outside of the defined threshold.

INSUFFICIENT_DATA - The alarm has just started, the metric is not available, or not enough data is available for the metric to determine the alarm state.

OK - The metric is within the defined threshold.

When you specify a notification trigger, the ALARM state must be specified. The INSUFFICIENT_DATA and OK states can be specified in addition to the ALARM state.

If you specify OK as an alarm trigger, a notification is sent when the alarm switches from an ALARM or INSUFFICIENT_DATA alarm state to an OK state. This can be thought of as an all clear alarm notification.

If you specify INSUFFICIENT_DATA as the alarm trigger, a notification is sent when the alarm switches from an OK or ALARM alarm state to an INSUFFICIENT_DATA state.

The notification trigger defaults to ALARM if you don't specify this parameter.

$sel:alarmName:PutAlarm', putAlarm_alarmName - The name for the alarm. Specify the name of an existing alarm to update, and overwrite the previous configuration of the alarm.

$sel:metricName:PutAlarm', putAlarm_metricName - The name of the metric to associate with the alarm.

You can configure up to two alarms per metric.

The following metrics are available for each resource type:

Instances: BurstCapacityPercentage, BurstCapacityTime, CPUUtilization, NetworkIn, NetworkOut, StatusCheckFailed, StatusCheckFailed_Instance, and StatusCheckFailed_System.

Load balancers: ClientTLSNegotiationErrorCount, HealthyHostCount, UnhealthyHostCount, HTTPCode_LB_4XX_Count, HTTPCode_LB_5XX_Count, HTTPCode_Instance_2XX_Count, HTTPCode_Instance_3XX_Count, HTTPCode_Instance_4XX_Count, HTTPCode_Instance_5XX_Count, InstanceResponseTime, RejectedConnectionCount, and RequestCount.

Relational databases: CPUUtilization, DatabaseConnections, DiskQueueDepth, FreeStorageSpace, NetworkReceiveThroughput, and NetworkTransmitThroughput.

For more information about these metrics, see Metrics available in Lightsail.

$sel:monitoredResourceName:PutAlarm', putAlarm_monitoredResourceName - The name of the Lightsail resource that will be monitored.

Instances, load balancers, and relational databases are the only Lightsail resources that can currently be monitored by alarms.

$sel:comparisonOperator:PutAlarm', putAlarm_comparisonOperator - The arithmetic operation to use when comparing the specified statistic to the threshold. The specified statistic value is used as the first operand.

$sel:threshold:PutAlarm', putAlarm_threshold - The value against which the specified statistic is compared.

$sel:evaluationPeriods:PutAlarm', putAlarm_evaluationPeriods - The number of most recent periods over which data is compared to the specified threshold. If you are setting an "M out of N" alarm, this value (evaluationPeriods) is the N.

If you are setting an alarm that requires that a number of consecutive data points be breaching to trigger the alarm, this value specifies the rolling period of time in which data points are evaluated.

Each evaluation period is five minutes long. For example, specify an evaluation period of 24 to evaluate a metric over a rolling period of two hours.

You can specify a minimum valuation period of 1 (5 minutes), and a maximum evaluation period of 288 (24 hours).

Request Lenses

putAlarm_treatMissingData :: Lens' PutAlarm (Maybe TreatMissingData)

Sets how this alarm will handle missing data points.

An alarm can treat missing data in the following ways:

breaching - Assume the missing data is not within the threshold. Missing data counts towards the number of times the metric is not within the threshold.

notBreaching - Assume the missing data is within the threshold. Missing data does not count towards the number of times the metric is not within the threshold.

ignore - Ignore the missing data. Maintains the current alarm state.

missing - Missing data is treated as missing.

If treatMissingData is not specified, the default behavior of missing is used.

putAlarm_contactProtocols :: Lens' PutAlarm (Maybe [ContactProtocol])

The contact protocols to use for the alarm, such as Email, SMS (text messaging), or both.

A notification is sent via the specified contact protocol if notifications are enabled for the alarm, and when the alarm is triggered.

A notification is not sent if a contact protocol is not specified, if the specified contact protocol is not configured in the AWS Region, or if notifications are not enabled for the alarm using the notificationEnabled paramater.

Use the CreateContactMethod action to configure a contact protocol in an AWS Region.

putAlarm_datapointsToAlarm :: Lens' PutAlarm (Maybe Int)

The number of data points that must be not within the specified threshold to trigger the alarm. If you are setting an "M out of N" alarm, this value (datapointsToAlarm) is the M.

putAlarm_notificationEnabled :: Lens' PutAlarm (Maybe Bool)

Indicates whether the alarm is enabled.

Notifications are enabled by default if you don't specify this parameter.

putAlarm_notificationTriggers :: Lens' PutAlarm (Maybe [AlarmState])

The alarm states that trigger a notification.

An alarm has the following possible states:

ALARM - The metric is outside of the defined threshold.

INSUFFICIENT_DATA - The alarm has just started, the metric is not available, or not enough data is available for the metric to determine the alarm state.

OK - The metric is within the defined threshold.

When you specify a notification trigger, the ALARM state must be specified. The INSUFFICIENT_DATA and OK states can be specified in addition to the ALARM state.

If you specify OK as an alarm trigger, a notification is sent when the alarm switches from an ALARM or INSUFFICIENT_DATA alarm state to an OK state. This can be thought of as an all clear alarm notification.

If you specify INSUFFICIENT_DATA as the alarm trigger, a notification is sent when the alarm switches from an OK or ALARM alarm state to an INSUFFICIENT_DATA state.

The notification trigger defaults to ALARM if you don't specify this parameter.

putAlarm_alarmName :: Lens' PutAlarm Text

The name for the alarm. Specify the name of an existing alarm to update, and overwrite the previous configuration of the alarm.

putAlarm_metricName :: Lens' PutAlarm MetricName

The name of the metric to associate with the alarm.

You can configure up to two alarms per metric.

The following metrics are available for each resource type:

Instances: BurstCapacityPercentage, BurstCapacityTime, CPUUtilization, NetworkIn, NetworkOut, StatusCheckFailed, StatusCheckFailed_Instance, and StatusCheckFailed_System.

Load balancers: ClientTLSNegotiationErrorCount, HealthyHostCount, UnhealthyHostCount, HTTPCode_LB_4XX_Count, HTTPCode_LB_5XX_Count, HTTPCode_Instance_2XX_Count, HTTPCode_Instance_3XX_Count, HTTPCode_Instance_4XX_Count, HTTPCode_Instance_5XX_Count, InstanceResponseTime, RejectedConnectionCount, and RequestCount.

Relational databases: CPUUtilization, DatabaseConnections, DiskQueueDepth, FreeStorageSpace, NetworkReceiveThroughput, and NetworkTransmitThroughput.

For more information about these metrics, see Metrics available in Lightsail.

putAlarm_monitoredResourceName :: Lens' PutAlarm Text

The name of the Lightsail resource that will be monitored.

Instances, load balancers, and relational databases are the only Lightsail resources that can currently be monitored by alarms.

putAlarm_comparisonOperator :: Lens' PutAlarm ComparisonOperator

The arithmetic operation to use when comparing the specified statistic to the threshold. The specified statistic value is used as the first operand.

putAlarm_threshold :: Lens' PutAlarm Double

The value against which the specified statistic is compared.

putAlarm_evaluationPeriods :: Lens' PutAlarm Int

The number of most recent periods over which data is compared to the specified threshold. If you are setting an "M out of N" alarm, this value (evaluationPeriods) is the N.

If you are setting an alarm that requires that a number of consecutive data points be breaching to trigger the alarm, this value specifies the rolling period of time in which data points are evaluated.

Each evaluation period is five minutes long. For example, specify an evaluation period of 24 to evaluate a metric over a rolling period of two hours.

You can specify a minimum valuation period of 1 (5 minutes), and a maximum evaluation period of 288 (24 hours).

Destructuring the Response

data PutAlarmResponse

See: newPutAlarmResponse smart constructor.

Constructors

PutAlarmResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newPutAlarmResponse

Arguments

:: Int
$sel:httpStatus:PutAlarmResponse'

-> PutAlarmResponse

Create a value of PutAlarmResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:PutAlarmResponse', putAlarmResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:PutAlarmResponse', putAlarmResponse_httpStatus - The response's http status code.

Response Lenses

putAlarmResponse_operations :: Lens' PutAlarmResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

putAlarmResponse_httpStatus :: Lens' PutAlarmResponse Int

The response's http status code.