Network.AWS.Lightsail.SendContactMethodVerification

Description

Sends a verification request to an email contact method to ensure it's owned by the requester. SMS contact methods don't need to be verified.

A contact method is used to send you notifications about your Amazon Lightsail resources. You can add one email address and one mobile phone number contact method in each AWS Region. However, SMS text messaging is not supported in some AWS Regions, and SMS text messages cannot be sent to some countries/regions. For more information, see Notifications in Amazon Lightsail.

A verification request is sent to the contact method when you initially create it. Use this action to send another verification request if a previous verification request was deleted, or has expired.

Notifications are not sent to an email contact method until after it is verified, and confirmed as valid.

Creating a Request

data SendContactMethodVerification

See: newSendContactMethodVerification smart constructor.

Constructors

SendContactMethodVerification'

Fields

protocol :: ContactMethodVerificationProtocol

The protocol to verify, such as Email or SMS (text messaging).

newSendContactMethodVerification

Arguments

:: ContactMethodVerificationProtocol
$sel:protocol:SendContactMethodVerification'

-> SendContactMethodVerification

Create a value of SendContactMethodVerification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protocol:SendContactMethodVerification', sendContactMethodVerification_protocol - The protocol to verify, such as Email or SMS (text messaging).

Request Lenses

sendContactMethodVerification_protocol :: Lens' SendContactMethodVerification ContactMethodVerificationProtocol

The protocol to verify, such as Email or SMS (text messaging).

Destructuring the Response

data SendContactMethodVerificationResponse

See: newSendContactMethodVerificationResponse smart constructor.

Constructors

SendContactMethodVerificationResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newSendContactMethodVerificationResponse

Arguments

:: Int
$sel:httpStatus:SendContactMethodVerificationResponse'

-> SendContactMethodVerificationResponse

Create a value of SendContactMethodVerificationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:SendContactMethodVerificationResponse', sendContactMethodVerificationResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:SendContactMethodVerificationResponse', sendContactMethodVerificationResponse_httpStatus - The response's http status code.

Response Lenses

sendContactMethodVerificationResponse_operations :: Lens' SendContactMethodVerificationResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

sendContactMethodVerificationResponse_httpStatus :: Lens' SendContactMethodVerificationResponse Int

The response's http status code.