Network.AWS.Lightsail.SetResourceAccessForBucket

Description

Sets the Amazon Lightsail resources that can access the specified Lightsail bucket.

Lightsail buckets currently support setting access for Lightsail instances in the same AWS Region.

Creating a Request

data SetResourceAccessForBucket

See: newSetResourceAccessForBucket smart constructor.

Constructors

SetResourceAccessForBucket'

Fields

resourceName :: Text

The name of the Lightsail instance for which to set bucket access. The instance must be in a running or stopped state.

bucketName :: Text

The name of the bucket for which to set access to another Lightsail resource.

access :: ResourceBucketAccess

The access setting.

The following access settings are available:

allow - Allows access to the bucket and its objects.

deny - Denies access to the bucket and its objects. Use this setting to remove access for a resource previously set to allow.

newSetResourceAccessForBucket

Arguments

:: Text
$sel:resourceName:SetResourceAccessForBucket'

-> Text
$sel:bucketName:SetResourceAccessForBucket'

-> ResourceBucketAccess
$sel:access:SetResourceAccessForBucket'

-> SetResourceAccessForBucket

Create a value of SetResourceAccessForBucket with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceName:SetResourceAccessForBucket', setResourceAccessForBucket_resourceName - The name of the Lightsail instance for which to set bucket access. The instance must be in a running or stopped state.

$sel:bucketName:SetResourceAccessForBucket', setResourceAccessForBucket_bucketName - The name of the bucket for which to set access to another Lightsail resource.

$sel:access:SetResourceAccessForBucket', setResourceAccessForBucket_access - The access setting.

The following access settings are available:

allow - Allows access to the bucket and its objects.

deny - Denies access to the bucket and its objects. Use this setting to remove access for a resource previously set to allow.

Request Lenses

setResourceAccessForBucket_resourceName :: Lens' SetResourceAccessForBucket Text

The name of the Lightsail instance for which to set bucket access. The instance must be in a running or stopped state.

setResourceAccessForBucket_bucketName :: Lens' SetResourceAccessForBucket Text

The name of the bucket for which to set access to another Lightsail resource.

setResourceAccessForBucket_access :: Lens' SetResourceAccessForBucket ResourceBucketAccess

The access setting.

The following access settings are available:

allow - Allows access to the bucket and its objects.

deny - Denies access to the bucket and its objects. Use this setting to remove access for a resource previously set to allow.

Destructuring the Response

data SetResourceAccessForBucketResponse

See: newSetResourceAccessForBucketResponse smart constructor.

Constructors

SetResourceAccessForBucketResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newSetResourceAccessForBucketResponse

Arguments

:: Int
$sel:httpStatus:SetResourceAccessForBucketResponse'

-> SetResourceAccessForBucketResponse

Create a value of SetResourceAccessForBucketResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:SetResourceAccessForBucketResponse', setResourceAccessForBucketResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:SetResourceAccessForBucketResponse', setResourceAccessForBucketResponse_httpStatus - The response's http status code.

Response Lenses

setResourceAccessForBucketResponse_operations :: Lens' SetResourceAccessForBucketResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

setResourceAccessForBucketResponse_httpStatus :: Lens' SetResourceAccessForBucketResponse Int

The response's http status code.