Network.AWS.Lightsail.Types.AccessKey

Description

Documentation

data AccessKey

Describes an access key for an Amazon Lightsail bucket.

Access keys grant full programmatic access to the specified bucket and its objects. You can have a maximum of two access keys per bucket. Use the CreateBucketAccessKey action to create an access key for a specific bucket. For more information about access keys, see Creating access keys for a bucket in Amazon Lightsail in the Amazon Lightsail Developer Guide.

The secretAccessKey value is returned only in response to the CreateBucketAccessKey action. You can get a secret access key only when you first create an access key; you cannot get the secret access key later. If you lose the secret access key, you must create a new access key.

See: newAccessKey smart constructor.

Constructors

AccessKey'

Fields

status :: Maybe StatusType

The status of the access key.

A status of Active means that the key is valid, while Inactive means it is not.

createdAt :: Maybe POSIX

The timestamp when the access key was created.

secretAccessKey :: Maybe Text

The secret access key used to sign requests.

You should store the secret access key in a safe location. We recommend that you delete the access key if the secret access key is compromised.

lastUsed :: Maybe AccessKeyLastUsed

An object that describes the last time the access key was used.

This object does not include data in the response of a CreateBucketAccessKey action. If the access key has not been used, the region and serviceName values are N/A, and the lastUsedDate value is null.

accessKeyId :: Maybe (Sensitive Text)

The ID of the access key.

newAccessKey :: AccessKey

Create a value of AccessKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:AccessKey', accessKey_status - The status of the access key.

A status of Active means that the key is valid, while Inactive means it is not.

$sel:createdAt:AccessKey', accessKey_createdAt - The timestamp when the access key was created.

$sel:secretAccessKey:AccessKey', accessKey_secretAccessKey - The secret access key used to sign requests.

You should store the secret access key in a safe location. We recommend that you delete the access key if the secret access key is compromised.

$sel:lastUsed:AccessKey', accessKey_lastUsed - An object that describes the last time the access key was used.

This object does not include data in the response of a CreateBucketAccessKey action. If the access key has not been used, the region and serviceName values are N/A, and the lastUsedDate value is null.

$sel:accessKeyId:AccessKey', accessKey_accessKeyId - The ID of the access key.

accessKey_status :: Lens' AccessKey (Maybe StatusType)

The status of the access key.

A status of Active means that the key is valid, while Inactive means it is not.

accessKey_createdAt :: Lens' AccessKey (Maybe UTCTime)

The timestamp when the access key was created.

accessKey_secretAccessKey :: Lens' AccessKey (Maybe Text)

The secret access key used to sign requests.

You should store the secret access key in a safe location. We recommend that you delete the access key if the secret access key is compromised.

accessKey_lastUsed :: Lens' AccessKey (Maybe AccessKeyLastUsed)

An object that describes the last time the access key was used.

This object does not include data in the response of a CreateBucketAccessKey action. If the access key has not been used, the region and serviceName values are N/A, and the lastUsedDate value is null.

accessKey_accessKeyId :: Lens' AccessKey (Maybe Text)

The ID of the access key.