Network.AWS.Lightsail.Types.AddOn

Description

Documentation

data AddOn

Describes an add-on that is enabled for an Amazon Lightsail resource.

See: newAddOn smart constructor.

Constructors

AddOn'

Fields

snapshotTimeOfDay :: Maybe Text

The daily time when an automatic snapshot is created.

The time shown is in HH:00 format, and in Coordinated Universal Time (UTC).

The snapshot is automatically created between the time shown and up to 45 minutes after.

status :: Maybe Text

The status of the add-on.

name :: Maybe Text

The name of the add-on.

nextSnapshotTimeOfDay :: Maybe Text

The next daily time an automatic snapshot will be created.

The time shown is in HH:00 format, and in Coordinated Universal Time (UTC).

The snapshot is automatically created between the time shown and up to 45 minutes after.

newAddOn :: AddOn

Create a value of AddOn with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshotTimeOfDay:AddOn', addOn_snapshotTimeOfDay - The daily time when an automatic snapshot is created.

The time shown is in HH:00 format, and in Coordinated Universal Time (UTC).

The snapshot is automatically created between the time shown and up to 45 minutes after.

$sel:status:AddOn', addOn_status - The status of the add-on.

$sel:name:AddOn', addOn_name - The name of the add-on.

$sel:nextSnapshotTimeOfDay:AddOn', addOn_nextSnapshotTimeOfDay - The next daily time an automatic snapshot will be created.

The time shown is in HH:00 format, and in Coordinated Universal Time (UTC).

The snapshot is automatically created between the time shown and up to 45 minutes after.

addOn_snapshotTimeOfDay :: Lens' AddOn (Maybe Text)

The daily time when an automatic snapshot is created.

The time shown is in HH:00 format, and in Coordinated Universal Time (UTC).

The snapshot is automatically created between the time shown and up to 45 minutes after.

addOn_status :: Lens' AddOn (Maybe Text)

The status of the add-on.

addOn_name :: Lens' AddOn (Maybe Text)

The name of the add-on.

addOn_nextSnapshotTimeOfDay :: Lens' AddOn (Maybe Text)

The next daily time an automatic snapshot will be created.

The time shown is in HH:00 format, and in Coordinated Universal Time (UTC).

The snapshot is automatically created between the time shown and up to 45 minutes after.