Network.AWS.Lightsail.Types.Alarm

Description

Documentation

data Alarm

Describes an alarm.

An alarm is a way to monitor your Lightsail resource metrics. For more information, see Alarms in Amazon Lightsail.

See: newAlarm smart constructor.

Constructors

Alarm'

Fields

datapointsToAlarm :: Maybe Int

The number of data points that must not within the specified threshold to trigger the alarm.

threshold :: Maybe Double

The value against which the specified statistic is compared.

comparisonOperator :: Maybe ComparisonOperator

The arithmetic operation used when comparing the specified statistic and threshold.

monitoredResourceInfo :: Maybe MonitoredResourceInfo

An object that lists information about the resource monitored by the alarm.

unit :: Maybe MetricUnit

The unit of the metric associated with the alarm.

notificationTriggers :: Maybe [AlarmState]

The alarm states that trigger a notification.

metricName :: Maybe MetricName

The name of the metric associated with the alarm.

notificationEnabled :: Maybe Bool

Indicates whether the alarm is enabled.

createdAt :: Maybe POSIX

The timestamp when the alarm was created.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the alarm.

supportCode :: Maybe Text

The support code. Include this code in your email to support when you have questions about your Lightsail alarm. This code enables our support team to look up your Lightsail information more easily.

treatMissingData :: Maybe TreatMissingData

Specifies how the alarm handles missing data points.

An alarm can treat missing data in the following ways:

breaching - Assume the missing data is not within the threshold. Missing data counts towards the number of times the metric is not within the threshold.

notBreaching - Assume the missing data is within the threshold. Missing data does not count towards the number of times the metric is not within the threshold.

ignore - Ignore the missing data. Maintains the current alarm state.

missing - Missing data is treated as missing.

resourceType :: Maybe ResourceType

The Lightsail resource type (e.g., Alarm).

state :: Maybe AlarmState

The current state of the alarm.

An alarm has the following possible states:

ALARM - The metric is outside of the defined threshold.

INSUFFICIENT_DATA - The alarm has just started, the metric is not available, or not enough data is available for the metric to determine the alarm state.

OK - The metric is within the defined threshold.

name :: Maybe Text

The name of the alarm.

statistic :: Maybe MetricStatistic

The statistic for the metric associated with the alarm.

The following statistics are available:

Minimum - The lowest value observed during the specified period. Use this value to determine low volumes of activity for your application.

Maximum - The highest value observed during the specified period. Use this value to determine high volumes of activity for your application.

Sum - All values submitted for the matching metric added together. You can use this statistic to determine the total volume of a metric.

Average - The value of Sum / SampleCount during the specified period. By comparing this statistic with the Minimum and Maximum values, you can determine the full scope of a metric and how close the average use is to the Minimum and Maximum values. This comparison helps you to know when to increase or decrease your resources.

SampleCount - The count, or number, of data points used for the statistical calculation.

evaluationPeriods :: Maybe Int

The number of periods over which data is compared to the specified threshold.

contactProtocols :: Maybe [ContactProtocol]

The contact protocols for the alarm, such as Email, SMS (text messaging), or both.

period :: Maybe Natural

The period, in seconds, over which the statistic is applied.

location :: Maybe ResourceLocation

An object that lists information about the location of the alarm.

newAlarm :: Alarm

Create a value of Alarm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datapointsToAlarm:Alarm', alarm_datapointsToAlarm - The number of data points that must not within the specified threshold to trigger the alarm.

$sel:threshold:Alarm', alarm_threshold - The value against which the specified statistic is compared.

$sel:comparisonOperator:Alarm', alarm_comparisonOperator - The arithmetic operation used when comparing the specified statistic and threshold.

$sel:monitoredResourceInfo:Alarm', alarm_monitoredResourceInfo - An object that lists information about the resource monitored by the alarm.

$sel:unit:Alarm', alarm_unit - The unit of the metric associated with the alarm.

$sel:notificationTriggers:Alarm', alarm_notificationTriggers - The alarm states that trigger a notification.

$sel:metricName:Alarm', alarm_metricName - The name of the metric associated with the alarm.

$sel:notificationEnabled:Alarm', alarm_notificationEnabled - Indicates whether the alarm is enabled.

$sel:createdAt:Alarm', alarm_createdAt - The timestamp when the alarm was created.

$sel:arn:Alarm', alarm_arn - The Amazon Resource Name (ARN) of the alarm.

$sel:supportCode:Alarm', alarm_supportCode - The support code. Include this code in your email to support when you have questions about your Lightsail alarm. This code enables our support team to look up your Lightsail information more easily.

$sel:treatMissingData:Alarm', alarm_treatMissingData - Specifies how the alarm handles missing data points.

An alarm can treat missing data in the following ways:

breaching - Assume the missing data is not within the threshold. Missing data counts towards the number of times the metric is not within the threshold.

notBreaching - Assume the missing data is within the threshold. Missing data does not count towards the number of times the metric is not within the threshold.

ignore - Ignore the missing data. Maintains the current alarm state.

missing - Missing data is treated as missing.

$sel:resourceType:Alarm', alarm_resourceType - The Lightsail resource type (e.g., Alarm).

$sel:state:Alarm', alarm_state - The current state of the alarm.

An alarm has the following possible states:

ALARM - The metric is outside of the defined threshold.

INSUFFICIENT_DATA - The alarm has just started, the metric is not available, or not enough data is available for the metric to determine the alarm state.

OK - The metric is within the defined threshold.

$sel:name:Alarm', alarm_name - The name of the alarm.

$sel:statistic:Alarm', alarm_statistic - The statistic for the metric associated with the alarm.

The following statistics are available:

Minimum - The lowest value observed during the specified period. Use this value to determine low volumes of activity for your application.

Maximum - The highest value observed during the specified period. Use this value to determine high volumes of activity for your application.

Sum - All values submitted for the matching metric added together. You can use this statistic to determine the total volume of a metric.

Average - The value of Sum / SampleCount during the specified period. By comparing this statistic with the Minimum and Maximum values, you can determine the full scope of a metric and how close the average use is to the Minimum and Maximum values. This comparison helps you to know when to increase or decrease your resources.

SampleCount - The count, or number, of data points used for the statistical calculation.

$sel:evaluationPeriods:Alarm', alarm_evaluationPeriods - The number of periods over which data is compared to the specified threshold.

$sel:contactProtocols:Alarm', alarm_contactProtocols - The contact protocols for the alarm, such as Email, SMS (text messaging), or both.

$sel:period:Alarm', alarm_period - The period, in seconds, over which the statistic is applied.

$sel:location:Alarm', alarm_location - An object that lists information about the location of the alarm.

alarm_datapointsToAlarm :: Lens' Alarm (Maybe Int)

The number of data points that must not within the specified threshold to trigger the alarm.

alarm_threshold :: Lens' Alarm (Maybe Double)

The value against which the specified statistic is compared.

alarm_comparisonOperator :: Lens' Alarm (Maybe ComparisonOperator)

The arithmetic operation used when comparing the specified statistic and threshold.

alarm_monitoredResourceInfo :: Lens' Alarm (Maybe MonitoredResourceInfo)

An object that lists information about the resource monitored by the alarm.

alarm_unit :: Lens' Alarm (Maybe MetricUnit)

The unit of the metric associated with the alarm.

alarm_notificationTriggers :: Lens' Alarm (Maybe [AlarmState])

The alarm states that trigger a notification.

alarm_metricName :: Lens' Alarm (Maybe MetricName)

The name of the metric associated with the alarm.

alarm_notificationEnabled :: Lens' Alarm (Maybe Bool)

Indicates whether the alarm is enabled.

alarm_createdAt :: Lens' Alarm (Maybe UTCTime)

The timestamp when the alarm was created.

alarm_arn :: Lens' Alarm (Maybe Text)

The Amazon Resource Name (ARN) of the alarm.

alarm_supportCode :: Lens' Alarm (Maybe Text)

The support code. Include this code in your email to support when you have questions about your Lightsail alarm. This code enables our support team to look up your Lightsail information more easily.

alarm_treatMissingData :: Lens' Alarm (Maybe TreatMissingData)

Specifies how the alarm handles missing data points.

An alarm can treat missing data in the following ways:

breaching - Assume the missing data is not within the threshold. Missing data counts towards the number of times the metric is not within the threshold.

notBreaching - Assume the missing data is within the threshold. Missing data does not count towards the number of times the metric is not within the threshold.

ignore - Ignore the missing data. Maintains the current alarm state.

missing - Missing data is treated as missing.

alarm_resourceType :: Lens' Alarm (Maybe ResourceType)

The Lightsail resource type (e.g., Alarm).

alarm_state :: Lens' Alarm (Maybe AlarmState)

The current state of the alarm.

An alarm has the following possible states:

ALARM - The metric is outside of the defined threshold.

INSUFFICIENT_DATA - The alarm has just started, the metric is not available, or not enough data is available for the metric to determine the alarm state.

OK - The metric is within the defined threshold.

alarm_name :: Lens' Alarm (Maybe Text)

The name of the alarm.

alarm_statistic :: Lens' Alarm (Maybe MetricStatistic)

The statistic for the metric associated with the alarm.

The following statistics are available:

Minimum - The lowest value observed during the specified period. Use this value to determine low volumes of activity for your application.

Maximum - The highest value observed during the specified period. Use this value to determine high volumes of activity for your application.

Sum - All values submitted for the matching metric added together. You can use this statistic to determine the total volume of a metric.

Average - The value of Sum / SampleCount during the specified period. By comparing this statistic with the Minimum and Maximum values, you can determine the full scope of a metric and how close the average use is to the Minimum and Maximum values. This comparison helps you to know when to increase or decrease your resources.

SampleCount - The count, or number, of data points used for the statistical calculation.

alarm_evaluationPeriods :: Lens' Alarm (Maybe Int)

The number of periods over which data is compared to the specified threshold.

alarm_contactProtocols :: Lens' Alarm (Maybe [ContactProtocol])

The contact protocols for the alarm, such as Email, SMS (text messaging), or both.

alarm_period :: Lens' Alarm (Maybe Natural)

The period, in seconds, over which the statistic is applied.

alarm_location :: Lens' Alarm (Maybe ResourceLocation)

An object that lists information about the location of the alarm.