Network.AWS.Lightsail.Types.Blueprint

Description

Documentation

data Blueprint

Describes a blueprint (a virtual private server image).

See: newBlueprint smart constructor.

Constructors

Blueprint'

Fields

platform :: Maybe InstancePlatform

The operating system platform (either Linux/Unix-based or Windows Server-based) of the blueprint.

isActive :: Maybe Bool

A Boolean value indicating whether the blueprint is active. Inactive blueprints are listed to support customers with existing instances but are not necessarily available for launch of new instances. Blueprints are marked inactive when they become outdated due to operating system updates or new application releases.

licenseUrl :: Maybe Text

The end-user license agreement URL for the image or blueprint.

productUrl :: Maybe Text

The product URL to learn more about the image or blueprint.

blueprintId :: Maybe Text

The ID for the virtual private server image (e.g., app_wordpress_4_4 or app_lamp_7_0).

version :: Maybe Text

The version number of the operating system, application, or stack (e.g., 2016.03.0).

group' :: Maybe Text

The group name of the blueprint (e.g., amazon-linux).

name :: Maybe Text

The friendly name of the blueprint (e.g., Amazon Linux).

versionCode :: Maybe Text

The version code.

description :: Maybe Text

The description of the blueprint.

type' :: Maybe BlueprintType

The type of the blueprint (e.g., os or app).

minPower :: Maybe Int

The minimum bundle power required to run this blueprint. For example, you need a bundle with a power value of 500 or more to create an instance that uses a blueprint with a minimum power value of 500. 0 indicates that the blueprint runs on all instance sizes.

newBlueprint :: Blueprint

Create a value of Blueprint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platform:Blueprint', blueprint_platform - The operating system platform (either Linux/Unix-based or Windows Server-based) of the blueprint.

$sel:isActive:Blueprint', blueprint_isActive - A Boolean value indicating whether the blueprint is active. Inactive blueprints are listed to support customers with existing instances but are not necessarily available for launch of new instances. Blueprints are marked inactive when they become outdated due to operating system updates or new application releases.

$sel:licenseUrl:Blueprint', blueprint_licenseUrl - The end-user license agreement URL for the image or blueprint.

$sel:productUrl:Blueprint', blueprint_productUrl - The product URL to learn more about the image or blueprint.

$sel:blueprintId:Blueprint', blueprint_blueprintId - The ID for the virtual private server image (e.g., app_wordpress_4_4 or app_lamp_7_0).

$sel:version:Blueprint', blueprint_version - The version number of the operating system, application, or stack (e.g., 2016.03.0).

$sel:group':Blueprint', blueprint_group - The group name of the blueprint (e.g., amazon-linux).

$sel:name:Blueprint', blueprint_name - The friendly name of the blueprint (e.g., Amazon Linux).

$sel:versionCode:Blueprint', blueprint_versionCode - The version code.

$sel:description:Blueprint', blueprint_description - The description of the blueprint.

$sel:type':Blueprint', blueprint_type - The type of the blueprint (e.g., os or app).

$sel:minPower:Blueprint', blueprint_minPower - The minimum bundle power required to run this blueprint. For example, you need a bundle with a power value of 500 or more to create an instance that uses a blueprint with a minimum power value of 500. 0 indicates that the blueprint runs on all instance sizes.

blueprint_platform :: Lens' Blueprint (Maybe InstancePlatform)

The operating system platform (either Linux/Unix-based or Windows Server-based) of the blueprint.

blueprint_isActive :: Lens' Blueprint (Maybe Bool)

A Boolean value indicating whether the blueprint is active. Inactive blueprints are listed to support customers with existing instances but are not necessarily available for launch of new instances. Blueprints are marked inactive when they become outdated due to operating system updates or new application releases.

blueprint_licenseUrl :: Lens' Blueprint (Maybe Text)

The end-user license agreement URL for the image or blueprint.

blueprint_productUrl :: Lens' Blueprint (Maybe Text)

The product URL to learn more about the image or blueprint.

blueprint_blueprintId :: Lens' Blueprint (Maybe Text)

The ID for the virtual private server image (e.g., app_wordpress_4_4 or app_lamp_7_0).

blueprint_version :: Lens' Blueprint (Maybe Text)

The version number of the operating system, application, or stack (e.g., 2016.03.0).

blueprint_group :: Lens' Blueprint (Maybe Text)

The group name of the blueprint (e.g., amazon-linux).

blueprint_name :: Lens' Blueprint (Maybe Text)

The friendly name of the blueprint (e.g., Amazon Linux).

blueprint_versionCode :: Lens' Blueprint (Maybe Text)

The version code.

blueprint_description :: Lens' Blueprint (Maybe Text)

The description of the blueprint.

blueprint_type :: Lens' Blueprint (Maybe BlueprintType)

The type of the blueprint (e.g., os or app).

blueprint_minPower :: Lens' Blueprint (Maybe Int)

The minimum bundle power required to run this blueprint. For example, you need a bundle with a power value of 500 or more to create an instance that uses a blueprint with a minimum power value of 500. 0 indicates that the blueprint runs on all instance sizes.