Network.AWS.Lightsail.Types.BucketBundle

Description

Documentation

data BucketBundle

Describes the specifications of a bundle that can be applied to an Amazon Lightsail bucket.

A bucket bundle specifies the monthly cost, storage space, and data transfer quota for a bucket.

See: newBucketBundle smart constructor.

Constructors

BucketBundle'

Fields

bundleId :: Maybe Text

The ID of the bundle.

isActive :: Maybe Bool

Indicates whether the bundle is active. Use for a new or existing bucket.

name :: Maybe Text

The name of the bundle.

storagePerMonthInGb :: Maybe Int

The storage size of the bundle, in GB.

transferPerMonthInGb :: Maybe Int

The monthly network transfer quota of the bundle.

price :: Maybe Double

The monthly price of the bundle, in US dollars.

newBucketBundle :: BucketBundle

Create a value of BucketBundle with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bundleId:BucketBundle', bucketBundle_bundleId - The ID of the bundle.

$sel:isActive:BucketBundle', bucketBundle_isActive - Indicates whether the bundle is active. Use for a new or existing bucket.

$sel:name:BucketBundle', bucketBundle_name - The name of the bundle.

$sel:storagePerMonthInGb:BucketBundle', bucketBundle_storagePerMonthInGb - The storage size of the bundle, in GB.

$sel:transferPerMonthInGb:BucketBundle', bucketBundle_transferPerMonthInGb - The monthly network transfer quota of the bundle.

$sel:price:BucketBundle', bucketBundle_price - The monthly price of the bundle, in US dollars.

bucketBundle_bundleId :: Lens' BucketBundle (Maybe Text)

The ID of the bundle.

bucketBundle_isActive :: Lens' BucketBundle (Maybe Bool)

Indicates whether the bundle is active. Use for a new or existing bucket.

bucketBundle_name :: Lens' BucketBundle (Maybe Text)

The name of the bundle.

bucketBundle_storagePerMonthInGb :: Lens' BucketBundle (Maybe Int)

The storage size of the bundle, in GB.

bucketBundle_transferPerMonthInGb :: Lens' BucketBundle (Maybe Int)

The monthly network transfer quota of the bundle.

bucketBundle_price :: Lens' BucketBundle (Maybe Double)

The monthly price of the bundle, in US dollars.