Network.AWS.Lightsail.Types.Bundle

Description

Documentation

data Bundle

Describes a bundle, which is a set of specs describing your virtual private server (or instance).

See: newBundle smart constructor.

Constructors

Bundle'

Fields

power :: Maybe Int

A numeric value that represents the power of the bundle (e.g., 500). You can use the bundle's power value in conjunction with a blueprint's minimum power value to determine whether the blueprint will run on the bundle. For example, you need a bundle with a power value of 500 or more to create an instance that uses a blueprint with a minimum power value of 500.

ramSizeInGb :: Maybe Double

The amount of RAM in GB (e.g., 2.0).

instanceType :: Maybe Text

The Amazon EC2 instance type (e.g., t2.micro).

bundleId :: Maybe Text

The bundle ID (e.g., micro_1_0).

isActive :: Maybe Bool

A Boolean value indicating whether the bundle is active.

name :: Maybe Text

A friendly name for the bundle (e.g., Micro).

transferPerMonthInGb :: Maybe Int

The data transfer rate per month in GB (e.g., 2000).

cpuCount :: Maybe Int

The number of vCPUs included in the bundle (e.g., 2).

price :: Maybe Double

The price in US dollars (e.g., 5.0) of the bundle.

supportedPlatforms :: Maybe [InstancePlatform]

The operating system platform (Linux/Unix-based or Windows Server-based) that the bundle supports. You can only launch a WINDOWS bundle on a blueprint that supports the WINDOWS platform. LINUX_UNIX blueprints require a LINUX_UNIX bundle.

diskSizeInGb :: Maybe Int

The size of the SSD (e.g., 30).

newBundle :: Bundle

Create a value of Bundle with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:power:Bundle', bundle_power - A numeric value that represents the power of the bundle (e.g., 500). You can use the bundle's power value in conjunction with a blueprint's minimum power value to determine whether the blueprint will run on the bundle. For example, you need a bundle with a power value of 500 or more to create an instance that uses a blueprint with a minimum power value of 500.

$sel:ramSizeInGb:Bundle', bundle_ramSizeInGb - The amount of RAM in GB (e.g., 2.0).

$sel:instanceType:Bundle', bundle_instanceType - The Amazon EC2 instance type (e.g., t2.micro).

$sel:bundleId:Bundle', bundle_bundleId - The bundle ID (e.g., micro_1_0).

$sel:isActive:Bundle', bundle_isActive - A Boolean value indicating whether the bundle is active.

$sel:name:Bundle', bundle_name - A friendly name for the bundle (e.g., Micro).

$sel:transferPerMonthInGb:Bundle', bundle_transferPerMonthInGb - The data transfer rate per month in GB (e.g., 2000).

$sel:cpuCount:Bundle', bundle_cpuCount - The number of vCPUs included in the bundle (e.g., 2).

$sel:price:Bundle', bundle_price - The price in US dollars (e.g., 5.0) of the bundle.

$sel:supportedPlatforms:Bundle', bundle_supportedPlatforms - The operating system platform (Linux/Unix-based or Windows Server-based) that the bundle supports. You can only launch a WINDOWS bundle on a blueprint that supports the WINDOWS platform. LINUX_UNIX blueprints require a LINUX_UNIX bundle.

$sel:diskSizeInGb:Bundle', bundle_diskSizeInGb - The size of the SSD (e.g., 30).

bundle_power :: Lens' Bundle (Maybe Int)

A numeric value that represents the power of the bundle (e.g., 500). You can use the bundle's power value in conjunction with a blueprint's minimum power value to determine whether the blueprint will run on the bundle. For example, you need a bundle with a power value of 500 or more to create an instance that uses a blueprint with a minimum power value of 500.

bundle_ramSizeInGb :: Lens' Bundle (Maybe Double)

The amount of RAM in GB (e.g., 2.0).

bundle_instanceType :: Lens' Bundle (Maybe Text)

The Amazon EC2 instance type (e.g., t2.micro).

bundle_bundleId :: Lens' Bundle (Maybe Text)

The bundle ID (e.g., micro_1_0).

bundle_isActive :: Lens' Bundle (Maybe Bool)

A Boolean value indicating whether the bundle is active.

bundle_name :: Lens' Bundle (Maybe Text)

A friendly name for the bundle (e.g., Micro).

bundle_transferPerMonthInGb :: Lens' Bundle (Maybe Int)

The data transfer rate per month in GB (e.g., 2000).

bundle_cpuCount :: Lens' Bundle (Maybe Int)

The number of vCPUs included in the bundle (e.g., 2).

bundle_price :: Lens' Bundle (Maybe Double)

The price in US dollars (e.g., 5.0) of the bundle.

bundle_supportedPlatforms :: Lens' Bundle (Maybe [InstancePlatform])

The operating system platform (Linux/Unix-based or Windows Server-based) that the bundle supports. You can only launch a WINDOWS bundle on a blueprint that supports the WINDOWS platform. LINUX_UNIX blueprints require a LINUX_UNIX bundle.

bundle_diskSizeInGb :: Lens' Bundle (Maybe Int)

The size of the SSD (e.g., 30).