Network.AWS.Lightsail.Types.Certificate

Description

Documentation

data Certificate

Describes the full details of an Amazon Lightsail SSL/TLS certificate.

To get a summary of a certificate, use the GetCertificates action and ommit includeCertificateDetails from your request. The response will include only the certificate Amazon Resource Name (ARN), certificate name, domain name, and tags.

See: newCertificate smart constructor.

Constructors

Certificate'

Fields

notBefore :: Maybe POSIX

The timestamp when the certificate is first valid.

status :: Maybe CertificateStatus

The validation status of the certificate.

eligibleToRenew :: Maybe Text

The renewal eligibility of the certificate.

createdAt :: Maybe POSIX

The timestamp when the certificate was created.

inUseResourceCount :: Maybe Int

The number of Lightsail resources that the certificate is attached to.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the certificate.

supportCode :: Maybe Text

The support code. Include this code in your email to support when you have questions about your Lightsail certificate. This code enables our support team to look up your Lightsail information more easily.

domainName :: Maybe Text

The domain name of the certificate.

subjectAlternativeNames :: Maybe [Text]

An array of strings that specify the alternate domains (e.g., example2.com) and subdomains (e.g., blog.example.com) of the certificate.

name :: Maybe Text

The name of the certificate (e.g., my-certificate).

revocationReason :: Maybe Text

The reason the certificate was revoked. This value is present only when the certificate status is REVOKED.

notAfter :: Maybe POSIX

The timestamp when the certificate expires.

issuerCA :: Maybe Text

The certificate authority that issued the certificate.

revokedAt :: Maybe POSIX

The timestamp when the certificate was revoked. This value is present only when the certificate status is REVOKED.

tags :: Maybe [Tag]

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

serialNumber :: Maybe Text

The serial number of the certificate.

keyAlgorithm :: Maybe Text

The algorithm used to generate the key pair (the public and private key) of the certificate.

issuedAt :: Maybe POSIX

The timestamp when the certificate was issued.

requestFailureReason :: Maybe Text

The validation failure reason, if any, of the certificate.

The following failure reasons are possible:

NO_AVAILABLE_CONTACTS - This failure applies to email validation, which is not available for Lightsail certificates.

ADDITIONAL_VERIFICATION_REQUIRED - Lightsail requires additional information to process this certificate request. This can happen as a fraud-protection measure, such as when the domain ranks within the Alexa top 1000 websites. To provide the required information, use the AWS Support Center to contact AWS Support.

You cannot request a certificate for Amazon-owned domain names such as those ending in amazonaws.com, cloudfront.net, or elasticbeanstalk.com.

DOMAIN_NOT_ALLOWED - One or more of the domain names in the certificate request was reported as an unsafe domain by VirusTotal. To correct the problem, search for your domain name on the VirusTotal website. If your domain is reported as suspicious, see Google Help for Hacked Websites to learn what you can do.

If you believe that the result is a false positive, notify the organization that is reporting the domain. VirusTotal is an aggregate of several antivirus and URL scanners and cannot remove your domain from a block list itself. After you correct the problem and the VirusTotal registry has been updated, request a new certificate.

If you see this error and your domain is not included in the VirusTotal list, visit the AWS Support Center and create a case.

INVALID_PUBLIC_DOMAIN - One or more of the domain names in the certificate request is not valid. Typically, this is because a domain name in the request is not a valid top-level domain. Try to request a certificate again, correcting any spelling errors or typos that were in the failed request, and ensure that all domain names in the request are for valid top-level domains. For example, you cannot request a certificate for example.invalidpublicdomain because invalidpublicdomain is not a valid top-level domain.

OTHER - Typically, this failure occurs when there is a typographical error in one or more of the domain names in the certificate request. Try to request a certificate again, correcting any spelling errors or typos that were in the failed request.

domainValidationRecords :: Maybe [DomainValidationRecord]

An array of objects that describe the domain validation records of the certificate.

renewalSummary :: Maybe RenewalSummary

An object that describes the status of the certificate renewal managed by Lightsail.

newCertificate :: Certificate

Create a value of Certificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:notBefore:Certificate', certificate_notBefore - The timestamp when the certificate is first valid.

$sel:status:Certificate', certificate_status - The validation status of the certificate.

$sel:eligibleToRenew:Certificate', certificate_eligibleToRenew - The renewal eligibility of the certificate.

$sel:createdAt:Certificate', certificate_createdAt - The timestamp when the certificate was created.

$sel:inUseResourceCount:Certificate', certificate_inUseResourceCount - The number of Lightsail resources that the certificate is attached to.

$sel:arn:Certificate', certificate_arn - The Amazon Resource Name (ARN) of the certificate.

$sel:supportCode:Certificate', certificate_supportCode - The support code. Include this code in your email to support when you have questions about your Lightsail certificate. This code enables our support team to look up your Lightsail information more easily.

$sel:domainName:Certificate', certificate_domainName - The domain name of the certificate.

$sel:subjectAlternativeNames:Certificate', certificate_subjectAlternativeNames - An array of strings that specify the alternate domains (e.g., example2.com) and subdomains (e.g., blog.example.com) of the certificate.

$sel:name:Certificate', certificate_name - The name of the certificate (e.g., my-certificate).

$sel:revocationReason:Certificate', certificate_revocationReason - The reason the certificate was revoked. This value is present only when the certificate status is REVOKED.

$sel:notAfter:Certificate', certificate_notAfter - The timestamp when the certificate expires.

$sel:issuerCA:Certificate', certificate_issuerCA - The certificate authority that issued the certificate.

$sel:revokedAt:Certificate', certificate_revokedAt - The timestamp when the certificate was revoked. This value is present only when the certificate status is REVOKED.

$sel:tags:Certificate', certificate_tags - The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

$sel:serialNumber:Certificate', certificate_serialNumber - The serial number of the certificate.

$sel:keyAlgorithm:Certificate', certificate_keyAlgorithm - The algorithm used to generate the key pair (the public and private key) of the certificate.

$sel:issuedAt:Certificate', certificate_issuedAt - The timestamp when the certificate was issued.

$sel:requestFailureReason:Certificate', certificate_requestFailureReason - The validation failure reason, if any, of the certificate.

The following failure reasons are possible:

NO_AVAILABLE_CONTACTS - This failure applies to email validation, which is not available for Lightsail certificates.

ADDITIONAL_VERIFICATION_REQUIRED - Lightsail requires additional information to process this certificate request. This can happen as a fraud-protection measure, such as when the domain ranks within the Alexa top 1000 websites. To provide the required information, use the AWS Support Center to contact AWS Support.

You cannot request a certificate for Amazon-owned domain names such as those ending in amazonaws.com, cloudfront.net, or elasticbeanstalk.com.

DOMAIN_NOT_ALLOWED - One or more of the domain names in the certificate request was reported as an unsafe domain by VirusTotal. To correct the problem, search for your domain name on the VirusTotal website. If your domain is reported as suspicious, see Google Help for Hacked Websites to learn what you can do.

If you believe that the result is a false positive, notify the organization that is reporting the domain. VirusTotal is an aggregate of several antivirus and URL scanners and cannot remove your domain from a block list itself. After you correct the problem and the VirusTotal registry has been updated, request a new certificate.

If you see this error and your domain is not included in the VirusTotal list, visit the AWS Support Center and create a case.

INVALID_PUBLIC_DOMAIN - One or more of the domain names in the certificate request is not valid. Typically, this is because a domain name in the request is not a valid top-level domain. Try to request a certificate again, correcting any spelling errors or typos that were in the failed request, and ensure that all domain names in the request are for valid top-level domains. For example, you cannot request a certificate for example.invalidpublicdomain because invalidpublicdomain is not a valid top-level domain.

OTHER - Typically, this failure occurs when there is a typographical error in one or more of the domain names in the certificate request. Try to request a certificate again, correcting any spelling errors or typos that were in the failed request.

$sel:domainValidationRecords:Certificate', certificate_domainValidationRecords - An array of objects that describe the domain validation records of the certificate.

$sel:renewalSummary:Certificate', certificate_renewalSummary - An object that describes the status of the certificate renewal managed by Lightsail.

certificate_notBefore :: Lens' Certificate (Maybe UTCTime)

The timestamp when the certificate is first valid.

certificate_status :: Lens' Certificate (Maybe CertificateStatus)

The validation status of the certificate.

certificate_eligibleToRenew :: Lens' Certificate (Maybe Text)

The renewal eligibility of the certificate.

certificate_createdAt :: Lens' Certificate (Maybe UTCTime)

The timestamp when the certificate was created.

certificate_inUseResourceCount :: Lens' Certificate (Maybe Int)

The number of Lightsail resources that the certificate is attached to.

certificate_arn :: Lens' Certificate (Maybe Text)

The Amazon Resource Name (ARN) of the certificate.

certificate_supportCode :: Lens' Certificate (Maybe Text)

The support code. Include this code in your email to support when you have questions about your Lightsail certificate. This code enables our support team to look up your Lightsail information more easily.

certificate_domainName :: Lens' Certificate (Maybe Text)

The domain name of the certificate.

certificate_subjectAlternativeNames :: Lens' Certificate (Maybe [Text])

An array of strings that specify the alternate domains (e.g., example2.com) and subdomains (e.g., blog.example.com) of the certificate.

certificate_name :: Lens' Certificate (Maybe Text)

The name of the certificate (e.g., my-certificate).

certificate_revocationReason :: Lens' Certificate (Maybe Text)

The reason the certificate was revoked. This value is present only when the certificate status is REVOKED.

certificate_notAfter :: Lens' Certificate (Maybe UTCTime)

The timestamp when the certificate expires.

certificate_issuerCA :: Lens' Certificate (Maybe Text)

The certificate authority that issued the certificate.

certificate_revokedAt :: Lens' Certificate (Maybe UTCTime)

The timestamp when the certificate was revoked. This value is present only when the certificate status is REVOKED.

certificate_tags :: Lens' Certificate (Maybe [Tag])

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

certificate_serialNumber :: Lens' Certificate (Maybe Text)

The serial number of the certificate.

certificate_keyAlgorithm :: Lens' Certificate (Maybe Text)

The algorithm used to generate the key pair (the public and private key) of the certificate.

certificate_issuedAt :: Lens' Certificate (Maybe UTCTime)

The timestamp when the certificate was issued.

certificate_requestFailureReason :: Lens' Certificate (Maybe Text)

The validation failure reason, if any, of the certificate.

The following failure reasons are possible:

NO_AVAILABLE_CONTACTS - This failure applies to email validation, which is not available for Lightsail certificates.

ADDITIONAL_VERIFICATION_REQUIRED - Lightsail requires additional information to process this certificate request. This can happen as a fraud-protection measure, such as when the domain ranks within the Alexa top 1000 websites. To provide the required information, use the AWS Support Center to contact AWS Support.

You cannot request a certificate for Amazon-owned domain names such as those ending in amazonaws.com, cloudfront.net, or elasticbeanstalk.com.

DOMAIN_NOT_ALLOWED - One or more of the domain names in the certificate request was reported as an unsafe domain by VirusTotal. To correct the problem, search for your domain name on the VirusTotal website. If your domain is reported as suspicious, see Google Help for Hacked Websites to learn what you can do.

If you believe that the result is a false positive, notify the organization that is reporting the domain. VirusTotal is an aggregate of several antivirus and URL scanners and cannot remove your domain from a block list itself. After you correct the problem and the VirusTotal registry has been updated, request a new certificate.

If you see this error and your domain is not included in the VirusTotal list, visit the AWS Support Center and create a case.

INVALID_PUBLIC_DOMAIN - One or more of the domain names in the certificate request is not valid. Typically, this is because a domain name in the request is not a valid top-level domain. Try to request a certificate again, correcting any spelling errors or typos that were in the failed request, and ensure that all domain names in the request are for valid top-level domains. For example, you cannot request a certificate for example.invalidpublicdomain because invalidpublicdomain is not a valid top-level domain.

OTHER - Typically, this failure occurs when there is a typographical error in one or more of the domain names in the certificate request. Try to request a certificate again, correcting any spelling errors or typos that were in the failed request.

certificate_domainValidationRecords :: Lens' Certificate (Maybe [DomainValidationRecord])

An array of objects that describe the domain validation records of the certificate.

certificate_renewalSummary :: Lens' Certificate (Maybe RenewalSummary)

An object that describes the status of the certificate renewal managed by Lightsail.