Network.AWS.Lightsail.Types.CloudFormationStackRecord

Description

Documentation

data CloudFormationStackRecord

Describes a CloudFormation stack record created as a result of the create cloud formation stack action.

A CloudFormation stack record provides information about the AWS CloudFormation stack used to create a new Amazon Elastic Compute Cloud instance from an exported Lightsail instance snapshot.

See: newCloudFormationStackRecord smart constructor.

Constructors

CloudFormationStackRecord'

Fields

state :: Maybe RecordState

The current state of the CloudFormation stack record.

destinationInfo :: Maybe DestinationInfo

A list of objects describing the destination service, which is AWS CloudFormation, and the Amazon Resource Name (ARN) of the AWS CloudFormation stack.

resourceType :: Maybe ResourceType

The Lightsail resource type (e.g., CloudFormationStackRecord).

arn :: Maybe Text

The Amazon Resource Name (ARN) of the CloudFormation stack record.

createdAt :: Maybe POSIX

The date when the CloudFormation stack record was created.

location :: Maybe ResourceLocation

A list of objects describing the Availability Zone and AWS Region of the CloudFormation stack record.

name :: Maybe Text

The name of the CloudFormation stack record. It starts with CloudFormationStackRecord followed by a GUID.

sourceInfo :: Maybe [CloudFormationStackRecordSourceInfo]

A list of objects describing the source of the CloudFormation stack record.

newCloudFormationStackRecord :: CloudFormationStackRecord

Create a value of CloudFormationStackRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:CloudFormationStackRecord', cloudFormationStackRecord_state - The current state of the CloudFormation stack record.

$sel:destinationInfo:CloudFormationStackRecord', cloudFormationStackRecord_destinationInfo - A list of objects describing the destination service, which is AWS CloudFormation, and the Amazon Resource Name (ARN) of the AWS CloudFormation stack.

$sel:resourceType:CloudFormationStackRecord', cloudFormationStackRecord_resourceType - The Lightsail resource type (e.g., CloudFormationStackRecord).

$sel:arn:CloudFormationStackRecord', cloudFormationStackRecord_arn - The Amazon Resource Name (ARN) of the CloudFormation stack record.

$sel:createdAt:CloudFormationStackRecord', cloudFormationStackRecord_createdAt - The date when the CloudFormation stack record was created.

$sel:location:CloudFormationStackRecord', cloudFormationStackRecord_location - A list of objects describing the Availability Zone and AWS Region of the CloudFormation stack record.

$sel:name:CloudFormationStackRecord', cloudFormationStackRecord_name - The name of the CloudFormation stack record. It starts with CloudFormationStackRecord followed by a GUID.

$sel:sourceInfo:CloudFormationStackRecord', cloudFormationStackRecord_sourceInfo - A list of objects describing the source of the CloudFormation stack record.

cloudFormationStackRecord_state :: Lens' CloudFormationStackRecord (Maybe RecordState)

The current state of the CloudFormation stack record.

cloudFormationStackRecord_destinationInfo :: Lens' CloudFormationStackRecord (Maybe DestinationInfo)

A list of objects describing the destination service, which is AWS CloudFormation, and the Amazon Resource Name (ARN) of the AWS CloudFormation stack.

cloudFormationStackRecord_resourceType :: Lens' CloudFormationStackRecord (Maybe ResourceType)

The Lightsail resource type (e.g., CloudFormationStackRecord).

cloudFormationStackRecord_arn :: Lens' CloudFormationStackRecord (Maybe Text)

The Amazon Resource Name (ARN) of the CloudFormation stack record.

cloudFormationStackRecord_createdAt :: Lens' CloudFormationStackRecord (Maybe UTCTime)

The date when the CloudFormation stack record was created.

cloudFormationStackRecord_location :: Lens' CloudFormationStackRecord (Maybe ResourceLocation)

A list of objects describing the Availability Zone and AWS Region of the CloudFormation stack record.

cloudFormationStackRecord_name :: Lens' CloudFormationStackRecord (Maybe Text)

The name of the CloudFormation stack record. It starts with CloudFormationStackRecord followed by a GUID.

cloudFormationStackRecord_sourceInfo :: Lens' CloudFormationStackRecord (Maybe [CloudFormationStackRecordSourceInfo])

A list of objects describing the source of the CloudFormation stack record.