Network.AWS.Lightsail.Types.ContainerService

Description

Documentation

data ContainerService

Describes an Amazon Lightsail container service.

See: newContainerService smart constructor.

Constructors

ContainerService'

Fields

power :: Maybe ContainerServicePowerName

The power specification of the container service.

The power specifies the amount of RAM, the number of vCPUs, and the base price of the container service.

stateDetail :: Maybe ContainerServiceStateDetail

An object that describes the current state of the container service.

The state detail is populated only when a container service is in a PENDING, DEPLOYING, or UPDATING state.

currentDeployment :: Maybe ContainerServiceDeployment

An object that describes the current container deployment of the container service.

createdAt :: Maybe POSIX

The timestamp when the container service was created.

containerServiceName :: Maybe Text

The name of the container service.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the container service.

privateDomainName :: Maybe Text

The private domain name of the container service.

The private domain name is accessible only by other resources within the default virtual private cloud (VPC) of your Lightsail account.

resourceType :: Maybe ResourceType

The Lightsail resource type of the container service (i.e., ContainerService).

state :: Maybe ContainerServiceState

The current state of the container service.

The following container service states are possible:

PENDING - The container service is being created.

READY - The container service is running but it does not have an active container deployment.

DEPLOYING - The container service is launching a container deployment.

RUNNING - The container service is running and it has an active container deployment.

UPDATING - The container service capacity or its custom domains are being updated.

DELETING - The container service is being deleted.

DISABLED - The container service is disabled, and its active deployment and containers, if any, are shut down.

principalArn :: Maybe Text

The principal ARN of the container service.

The principal ARN can be used to create a trust relationship between your standard AWS account and your Lightsail container service. This allows you to give your service permission to access resources in your standard AWS account.

tags :: Maybe [Tag]

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

nextDeployment :: Maybe ContainerServiceDeployment

An object that describes the next deployment of the container service.

This value is null when there is no deployment in a pending state.

url :: Maybe Text

The publicly accessible URL of the container service.

If no public endpoint is specified in the currentDeployment, this URL returns a 404 response.

scale :: Maybe Natural

The scale specification of the container service.

The scale specifies the allocated compute nodes of the container service.

publicDomainNames :: Maybe (HashMap Text [Text])

The public domain name of the container service, such as example.com and www.example.com.

You can specify up to four public domain names for a container service. The domain names that you specify are used when you create a deployment with a container configured as the public endpoint of your container service.

If you don't specify public domain names, then you can use the default domain of the container service.

You must create and validate an SSL/TLS certificate before you can use public domain names with your container service. Use the CreateCertificate action to create a certificate for the public domain names you want to use with your container service.

See CreateContainerService or UpdateContainerService for information about how to specify public domain names for your Lightsail container service.

location :: Maybe ResourceLocation

An object that describes the location of the container service, such as the AWS Region and Availability Zone.

powerId :: Maybe Text

The ID of the power of the container service.

isDisabled :: Maybe Bool

A Boolean value indicating whether the container service is disabled.

newContainerService :: ContainerService

Create a value of ContainerService with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:power:ContainerService', containerService_power - The power specification of the container service.

The power specifies the amount of RAM, the number of vCPUs, and the base price of the container service.

$sel:stateDetail:ContainerService', containerService_stateDetail - An object that describes the current state of the container service.

The state detail is populated only when a container service is in a PENDING, DEPLOYING, or UPDATING state.

$sel:currentDeployment:ContainerService', containerService_currentDeployment - An object that describes the current container deployment of the container service.

$sel:createdAt:ContainerService', containerService_createdAt - The timestamp when the container service was created.

$sel:containerServiceName:ContainerService', containerService_containerServiceName - The name of the container service.

$sel:arn:ContainerService', containerService_arn - The Amazon Resource Name (ARN) of the container service.

$sel:privateDomainName:ContainerService', containerService_privateDomainName - The private domain name of the container service.

The private domain name is accessible only by other resources within the default virtual private cloud (VPC) of your Lightsail account.

$sel:resourceType:ContainerService', containerService_resourceType - The Lightsail resource type of the container service (i.e., ContainerService).

$sel:state:ContainerService', containerService_state - The current state of the container service.

The following container service states are possible:

PENDING - The container service is being created.

READY - The container service is running but it does not have an active container deployment.

DEPLOYING - The container service is launching a container deployment.

RUNNING - The container service is running and it has an active container deployment.

UPDATING - The container service capacity or its custom domains are being updated.

DELETING - The container service is being deleted.

DISABLED - The container service is disabled, and its active deployment and containers, if any, are shut down.

$sel:principalArn:ContainerService', containerService_principalArn - The principal ARN of the container service.

The principal ARN can be used to create a trust relationship between your standard AWS account and your Lightsail container service. This allows you to give your service permission to access resources in your standard AWS account.

$sel:tags:ContainerService', containerService_tags - The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

$sel:nextDeployment:ContainerService', containerService_nextDeployment - An object that describes the next deployment of the container service.

This value is null when there is no deployment in a pending state.

$sel:url:ContainerService', containerService_url - The publicly accessible URL of the container service.

If no public endpoint is specified in the currentDeployment, this URL returns a 404 response.

$sel:scale:ContainerService', containerService_scale - The scale specification of the container service.

The scale specifies the allocated compute nodes of the container service.

$sel:publicDomainNames:ContainerService', containerService_publicDomainNames - The public domain name of the container service, such as example.com and www.example.com.

You can specify up to four public domain names for a container service. The domain names that you specify are used when you create a deployment with a container configured as the public endpoint of your container service.

If you don't specify public domain names, then you can use the default domain of the container service.

You must create and validate an SSL/TLS certificate before you can use public domain names with your container service. Use the CreateCertificate action to create a certificate for the public domain names you want to use with your container service.

See CreateContainerService or UpdateContainerService for information about how to specify public domain names for your Lightsail container service.

$sel:location:ContainerService', containerService_location - An object that describes the location of the container service, such as the AWS Region and Availability Zone.

$sel:powerId:ContainerService', containerService_powerId - The ID of the power of the container service.

$sel:isDisabled:ContainerService', containerService_isDisabled - A Boolean value indicating whether the container service is disabled.

containerService_power :: Lens' ContainerService (Maybe ContainerServicePowerName)

The power specification of the container service.

The power specifies the amount of RAM, the number of vCPUs, and the base price of the container service.

containerService_stateDetail :: Lens' ContainerService (Maybe ContainerServiceStateDetail)

An object that describes the current state of the container service.

The state detail is populated only when a container service is in a PENDING, DEPLOYING, or UPDATING state.

containerService_currentDeployment :: Lens' ContainerService (Maybe ContainerServiceDeployment)

An object that describes the current container deployment of the container service.

containerService_createdAt :: Lens' ContainerService (Maybe UTCTime)

The timestamp when the container service was created.

containerService_containerServiceName :: Lens' ContainerService (Maybe Text)

The name of the container service.

containerService_arn :: Lens' ContainerService (Maybe Text)

The Amazon Resource Name (ARN) of the container service.

containerService_privateDomainName :: Lens' ContainerService (Maybe Text)

The private domain name of the container service.

The private domain name is accessible only by other resources within the default virtual private cloud (VPC) of your Lightsail account.

containerService_resourceType :: Lens' ContainerService (Maybe ResourceType)

The Lightsail resource type of the container service (i.e., ContainerService).

containerService_state :: Lens' ContainerService (Maybe ContainerServiceState)

The current state of the container service.

The following container service states are possible:

PENDING - The container service is being created.

READY - The container service is running but it does not have an active container deployment.

DEPLOYING - The container service is launching a container deployment.

RUNNING - The container service is running and it has an active container deployment.

UPDATING - The container service capacity or its custom domains are being updated.

DELETING - The container service is being deleted.

DISABLED - The container service is disabled, and its active deployment and containers, if any, are shut down.

containerService_principalArn :: Lens' ContainerService (Maybe Text)

The principal ARN of the container service.

The principal ARN can be used to create a trust relationship between your standard AWS account and your Lightsail container service. This allows you to give your service permission to access resources in your standard AWS account.

containerService_tags :: Lens' ContainerService (Maybe [Tag])

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

containerService_nextDeployment :: Lens' ContainerService (Maybe ContainerServiceDeployment)

An object that describes the next deployment of the container service.

This value is null when there is no deployment in a pending state.

containerService_url :: Lens' ContainerService (Maybe Text)

The publicly accessible URL of the container service.

If no public endpoint is specified in the currentDeployment, this URL returns a 404 response.

containerService_scale :: Lens' ContainerService (Maybe Natural)

The scale specification of the container service.

The scale specifies the allocated compute nodes of the container service.

containerService_publicDomainNames :: Lens' ContainerService (Maybe (HashMap Text [Text]))

The public domain name of the container service, such as example.com and www.example.com.

You can specify up to four public domain names for a container service. The domain names that you specify are used when you create a deployment with a container configured as the public endpoint of your container service.

If you don't specify public domain names, then you can use the default domain of the container service.

You must create and validate an SSL/TLS certificate before you can use public domain names with your container service. Use the CreateCertificate action to create a certificate for the public domain names you want to use with your container service.

See CreateContainerService or UpdateContainerService for information about how to specify public domain names for your Lightsail container service.

containerService_location :: Lens' ContainerService (Maybe ResourceLocation)

An object that describes the location of the container service, such as the AWS Region and Availability Zone.

containerService_powerId :: Lens' ContainerService (Maybe Text)

The ID of the power of the container service.

containerService_isDisabled :: Lens' ContainerService (Maybe Bool)

A Boolean value indicating whether the container service is disabled.