Network.AWS.Lightsail.Types.ContainerServiceDeployment

Description

Documentation

data ContainerServiceDeployment

Describes a container deployment configuration of an Amazon Lightsail container service.

A deployment specifies the settings, such as the ports and launch command, of containers that are deployed to your container service.

See: newContainerServiceDeployment smart constructor.

Constructors

ContainerServiceDeployment'

Fields

state :: Maybe ContainerServiceDeploymentState

The state of the deployment.

A deployment can be in one of the following states:

Activating - The deployment is being created.

Active - The deployment was successfully created, and it's currently running on the container service. The container service can have only one deployment in an active state at a time.

Inactive - The deployment was previously successfully created, but it is not currently running on the container service.

Failed - The deployment failed. Use the GetContainerLog action to view the log events for the containers in the deployment to try to determine the reason for the failure.

publicEndpoint :: Maybe ContainerServiceEndpoint

An object that describes the endpoint of the deployment.

createdAt :: Maybe POSIX

The timestamp when the deployment was created.

containers :: Maybe (HashMap Text Container)

An object that describes the configuration for the containers of the deployment.

version :: Maybe Int

The version number of the deployment.

newContainerServiceDeployment :: ContainerServiceDeployment

Create a value of ContainerServiceDeployment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:ContainerServiceDeployment', containerServiceDeployment_state - The state of the deployment.

A deployment can be in one of the following states:

Activating - The deployment is being created.

Active - The deployment was successfully created, and it's currently running on the container service. The container service can have only one deployment in an active state at a time.

Inactive - The deployment was previously successfully created, but it is not currently running on the container service.

Failed - The deployment failed. Use the GetContainerLog action to view the log events for the containers in the deployment to try to determine the reason for the failure.

$sel:publicEndpoint:ContainerServiceDeployment', containerServiceDeployment_publicEndpoint - An object that describes the endpoint of the deployment.

$sel:createdAt:ContainerServiceDeployment', containerServiceDeployment_createdAt - The timestamp when the deployment was created.

$sel:containers:ContainerServiceDeployment', containerServiceDeployment_containers - An object that describes the configuration for the containers of the deployment.

$sel:version:ContainerServiceDeployment', containerServiceDeployment_version - The version number of the deployment.

containerServiceDeployment_state :: Lens' ContainerServiceDeployment (Maybe ContainerServiceDeploymentState)

The state of the deployment.

A deployment can be in one of the following states:

Activating - The deployment is being created.

Active - The deployment was successfully created, and it's currently running on the container service. The container service can have only one deployment in an active state at a time.

Inactive - The deployment was previously successfully created, but it is not currently running on the container service.

Failed - The deployment failed. Use the GetContainerLog action to view the log events for the containers in the deployment to try to determine the reason for the failure.

containerServiceDeployment_publicEndpoint :: Lens' ContainerServiceDeployment (Maybe ContainerServiceEndpoint)

An object that describes the endpoint of the deployment.

containerServiceDeployment_createdAt :: Lens' ContainerServiceDeployment (Maybe UTCTime)

The timestamp when the deployment was created.

containerServiceDeployment_containers :: Lens' ContainerServiceDeployment (Maybe (HashMap Text Container))

An object that describes the configuration for the containers of the deployment.

containerServiceDeployment_version :: Lens' ContainerServiceDeployment (Maybe Int)

The version number of the deployment.