Network.AWS.Lightsail.Types.ContainerServiceHealthCheckConfig

Description

Documentation

data ContainerServiceHealthCheckConfig

Describes the health check configuration of an Amazon Lightsail container service.

See: newContainerServiceHealthCheckConfig smart constructor.

Constructors

ContainerServiceHealthCheckConfig'

Fields

intervalSeconds :: Maybe Int

The approximate interval, in seconds, between health checks of an individual container. You can specify between 5 and 300 seconds. The default value is 5.

healthyThreshold :: Maybe Int

The number of consecutive health checks successes required before moving the container to the Healthy state. The default value is 2.

unhealthyThreshold :: Maybe Int

The number of consecutive health check failures required before moving the container to the Unhealthy state. The default value is 2.

timeoutSeconds :: Maybe Int

The amount of time, in seconds, during which no response means a failed health check. You can specify between 2 and 60 seconds. The default value is 2.

path :: Maybe Text

The path on the container on which to perform the health check. The default value is /.

successCodes :: Maybe Text

The HTTP codes to use when checking for a successful response from a container. You can specify values between 200 and 499. You can specify multiple values (for example, 200,202) or a range of values (for example, 200-299).

newContainerServiceHealthCheckConfig :: ContainerServiceHealthCheckConfig

Create a value of ContainerServiceHealthCheckConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:intervalSeconds:ContainerServiceHealthCheckConfig', containerServiceHealthCheckConfig_intervalSeconds - The approximate interval, in seconds, between health checks of an individual container. You can specify between 5 and 300 seconds. The default value is 5.

$sel:healthyThreshold:ContainerServiceHealthCheckConfig', containerServiceHealthCheckConfig_healthyThreshold - The number of consecutive health checks successes required before moving the container to the Healthy state. The default value is 2.

$sel:unhealthyThreshold:ContainerServiceHealthCheckConfig', containerServiceHealthCheckConfig_unhealthyThreshold - The number of consecutive health check failures required before moving the container to the Unhealthy state. The default value is 2.

$sel:timeoutSeconds:ContainerServiceHealthCheckConfig', containerServiceHealthCheckConfig_timeoutSeconds - The amount of time, in seconds, during which no response means a failed health check. You can specify between 2 and 60 seconds. The default value is 2.

$sel:path:ContainerServiceHealthCheckConfig', containerServiceHealthCheckConfig_path - The path on the container on which to perform the health check. The default value is /.

$sel:successCodes:ContainerServiceHealthCheckConfig', containerServiceHealthCheckConfig_successCodes - The HTTP codes to use when checking for a successful response from a container. You can specify values between 200 and 499. You can specify multiple values (for example, 200,202) or a range of values (for example, 200-299).

containerServiceHealthCheckConfig_intervalSeconds :: Lens' ContainerServiceHealthCheckConfig (Maybe Int)

The approximate interval, in seconds, between health checks of an individual container. You can specify between 5 and 300 seconds. The default value is 5.

containerServiceHealthCheckConfig_healthyThreshold :: Lens' ContainerServiceHealthCheckConfig (Maybe Int)

The number of consecutive health checks successes required before moving the container to the Healthy state. The default value is 2.

containerServiceHealthCheckConfig_unhealthyThreshold :: Lens' ContainerServiceHealthCheckConfig (Maybe Int)

The number of consecutive health check failures required before moving the container to the Unhealthy state. The default value is 2.

containerServiceHealthCheckConfig_timeoutSeconds :: Lens' ContainerServiceHealthCheckConfig (Maybe Int)

The amount of time, in seconds, during which no response means a failed health check. You can specify between 2 and 60 seconds. The default value is 2.

containerServiceHealthCheckConfig_path :: Lens' ContainerServiceHealthCheckConfig (Maybe Text)

The path on the container on which to perform the health check. The default value is /.

containerServiceHealthCheckConfig_successCodes :: Lens' ContainerServiceHealthCheckConfig (Maybe Text)

The HTTP codes to use when checking for a successful response from a container. You can specify values between 200 and 499. You can specify multiple values (for example, 200,202) or a range of values (for example, 200-299).