Network.AWS.Lightsail.Types.ContainerServicePower

Description

Documentation

data ContainerServicePower

Describes the powers that can be specified for an Amazon Lightsail container service.

The power specifies the amount of RAM, the number of vCPUs, and the base price of the container service.

See: newContainerServicePower smart constructor.

Constructors

ContainerServicePower'

Fields

powerId :: Maybe Text

The ID of the power (e.g., nano-1).

cpuCount :: Maybe Double

The number of vCPUs included in the power.

name :: Maybe Text

The friendly name of the power (e.g., nano).

price :: Maybe Double

The monthly price of the power in USD.

isActive :: Maybe Bool

A Boolean value indicating whether the power is active and can be specified for container services.

ramSizeInGb :: Maybe Double

The amount of RAM (in GB) of the power.

newContainerServicePower :: ContainerServicePower

Create a value of ContainerServicePower with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:powerId:ContainerServicePower', containerServicePower_powerId - The ID of the power (e.g., nano-1).

$sel:cpuCount:ContainerServicePower', containerServicePower_cpuCount - The number of vCPUs included in the power.

$sel:name:ContainerServicePower', containerServicePower_name - The friendly name of the power (e.g., nano).

$sel:price:ContainerServicePower', containerServicePower_price - The monthly price of the power in USD.

$sel:isActive:ContainerServicePower', containerServicePower_isActive - A Boolean value indicating whether the power is active and can be specified for container services.

$sel:ramSizeInGb:ContainerServicePower', containerServicePower_ramSizeInGb - The amount of RAM (in GB) of the power.

containerServicePower_powerId :: Lens' ContainerServicePower (Maybe Text)

The ID of the power (e.g., nano-1).

containerServicePower_cpuCount :: Lens' ContainerServicePower (Maybe Double)

The number of vCPUs included in the power.

containerServicePower_name :: Lens' ContainerServicePower (Maybe Text)

The friendly name of the power (e.g., nano).

containerServicePower_price :: Lens' ContainerServicePower (Maybe Double)

The monthly price of the power in USD.

containerServicePower_isActive :: Lens' ContainerServicePower (Maybe Bool)

A Boolean value indicating whether the power is active and can be specified for container services.

containerServicePower_ramSizeInGb :: Lens' ContainerServicePower (Maybe Double)

The amount of RAM (in GB) of the power.