Network.AWS.Lightsail.Types.Disk

Description

Documentation

data Disk

Describes a block storage disk.

See: newDisk smart constructor.

Constructors

Disk'

Fields

state :: Maybe DiskState

Describes the status of the disk.

resourceType :: Maybe ResourceType

The Lightsail resource type (e.g., Disk).

arn :: Maybe Text

The Amazon Resource Name (ARN) of the disk.

path :: Maybe Text

The disk path.

createdAt :: Maybe POSIX

The date when the disk was created.

location :: Maybe ResourceLocation

The AWS Region and Availability Zone where the disk is located.

iops :: Maybe Int

The input/output operations per second (IOPS) of the disk.

isAttached :: Maybe Bool

A Boolean value indicating whether the disk is attached.

addOns :: Maybe [AddOn]

An array of objects representing the add-ons enabled on the disk.

attachmentState :: Maybe Text

(Deprecated) The attachment state of the disk.

In releases prior to November 14, 2017, this parameter returned attached for system disks in the API response. It is now deprecated, but still included in the response. Use isAttached instead.

name :: Maybe Text

The unique name of the disk.

sizeInGb :: Maybe Int

The size of the disk in GB.

supportCode :: Maybe Text

The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

isSystemDisk :: Maybe Bool

A Boolean value indicating whether this disk is a system disk (has an operating system loaded on it).

attachedTo :: Maybe Text

The resources to which the disk is attached.

gbInUse :: Maybe Int

(Deprecated) The number of GB in use by the disk.

In releases prior to November 14, 2017, this parameter was not included in the API response. It is now deprecated.

tags :: Maybe [Tag]

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

newDisk :: Disk

Create a value of Disk with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:Disk', disk_state - Describes the status of the disk.

$sel:resourceType:Disk', disk_resourceType - The Lightsail resource type (e.g., Disk).

$sel:arn:Disk', disk_arn - The Amazon Resource Name (ARN) of the disk.

$sel:path:Disk', disk_path - The disk path.

$sel:createdAt:Disk', disk_createdAt - The date when the disk was created.

$sel:location:Disk', disk_location - The AWS Region and Availability Zone where the disk is located.

$sel:iops:Disk', disk_iops - The input/output operations per second (IOPS) of the disk.

$sel:isAttached:Disk', disk_isAttached - A Boolean value indicating whether the disk is attached.

$sel:addOns:Disk', disk_addOns - An array of objects representing the add-ons enabled on the disk.

$sel:attachmentState:Disk', disk_attachmentState - (Deprecated) The attachment state of the disk.

In releases prior to November 14, 2017, this parameter returned attached for system disks in the API response. It is now deprecated, but still included in the response. Use isAttached instead.

$sel:name:Disk', disk_name - The unique name of the disk.

$sel:sizeInGb:Disk', disk_sizeInGb - The size of the disk in GB.

$sel:supportCode:Disk', disk_supportCode - The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

$sel:isSystemDisk:Disk', disk_isSystemDisk - A Boolean value indicating whether this disk is a system disk (has an operating system loaded on it).

$sel:attachedTo:Disk', disk_attachedTo - The resources to which the disk is attached.

$sel:gbInUse:Disk', disk_gbInUse - (Deprecated) The number of GB in use by the disk.

In releases prior to November 14, 2017, this parameter was not included in the API response. It is now deprecated.

$sel:tags:Disk', disk_tags - The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

disk_state :: Lens' Disk (Maybe DiskState)

Describes the status of the disk.

disk_resourceType :: Lens' Disk (Maybe ResourceType)

The Lightsail resource type (e.g., Disk).

disk_arn :: Lens' Disk (Maybe Text)

The Amazon Resource Name (ARN) of the disk.

disk_path :: Lens' Disk (Maybe Text)

The disk path.

disk_createdAt :: Lens' Disk (Maybe UTCTime)

The date when the disk was created.

disk_location :: Lens' Disk (Maybe ResourceLocation)

The AWS Region and Availability Zone where the disk is located.

disk_iops :: Lens' Disk (Maybe Int)

The input/output operations per second (IOPS) of the disk.

disk_isAttached :: Lens' Disk (Maybe Bool)

A Boolean value indicating whether the disk is attached.

disk_addOns :: Lens' Disk (Maybe [AddOn])

An array of objects representing the add-ons enabled on the disk.

disk_attachmentState :: Lens' Disk (Maybe Text)

(Deprecated) The attachment state of the disk.

In releases prior to November 14, 2017, this parameter returned attached for system disks in the API response. It is now deprecated, but still included in the response. Use isAttached instead.

disk_name :: Lens' Disk (Maybe Text)

The unique name of the disk.

disk_sizeInGb :: Lens' Disk (Maybe Int)

The size of the disk in GB.

disk_supportCode :: Lens' Disk (Maybe Text)

The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

disk_isSystemDisk :: Lens' Disk (Maybe Bool)

A Boolean value indicating whether this disk is a system disk (has an operating system loaded on it).

disk_attachedTo :: Lens' Disk (Maybe Text)

The resources to which the disk is attached.

disk_gbInUse :: Lens' Disk (Maybe Int)

(Deprecated) The number of GB in use by the disk.

In releases prior to November 14, 2017, this parameter was not included in the API response. It is now deprecated.

disk_tags :: Lens' Disk (Maybe [Tag])

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.