Network.AWS.Lightsail.Types.Domain

Description

Documentation

data Domain

Describes a domain where you are storing recordsets.

See: newDomain smart constructor.

Constructors

Domain'

Fields

createdAt :: Maybe POSIX

The date when the domain recordset was created.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the domain recordset (e.g., arn:aws:lightsail:global:123456789101:Domain/824cede0-abc7-4f84-8dbc-12345EXAMPLE).

supportCode :: Maybe Text

The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

resourceType :: Maybe ResourceType

The resource type.

name :: Maybe Text

The name of the domain.

tags :: Maybe [Tag]

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

location :: Maybe ResourceLocation

The AWS Region and Availability Zones where the domain recordset was created.

domainEntries :: Maybe [DomainEntry]

An array of key-value pairs containing information about the domain entries.

newDomain :: Domain

Create a value of Domain with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdAt:Domain', domain_createdAt - The date when the domain recordset was created.

$sel:arn:Domain', domain_arn - The Amazon Resource Name (ARN) of the domain recordset (e.g., arn:aws:lightsail:global:123456789101:Domain/824cede0-abc7-4f84-8dbc-12345EXAMPLE).

$sel:supportCode:Domain', domain_supportCode - The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

$sel:resourceType:Domain', domain_resourceType - The resource type.

$sel:name:Domain', domain_name - The name of the domain.

$sel:tags:Domain', domain_tags - The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

$sel:location:Domain', domain_location - The AWS Region and Availability Zones where the domain recordset was created.

$sel:domainEntries:Domain', domain_domainEntries - An array of key-value pairs containing information about the domain entries.

domain_createdAt :: Lens' Domain (Maybe UTCTime)

The date when the domain recordset was created.

domain_arn :: Lens' Domain (Maybe Text)

The Amazon Resource Name (ARN) of the domain recordset (e.g., arn:aws:lightsail:global:123456789101:Domain/824cede0-abc7-4f84-8dbc-12345EXAMPLE).

domain_supportCode :: Lens' Domain (Maybe Text)

The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

domain_resourceType :: Lens' Domain (Maybe ResourceType)

The resource type.

domain_name :: Lens' Domain (Maybe Text)

The name of the domain.

domain_tags :: Lens' Domain (Maybe [Tag])

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

domain_location :: Lens' Domain (Maybe ResourceLocation)

The AWS Region and Availability Zones where the domain recordset was created.

domain_domainEntries :: Lens' Domain (Maybe [DomainEntry])

An array of key-value pairs containing information about the domain entries.