Network.AWS.Lightsail.Types.DomainValidationRecord

Description

Documentation

data DomainValidationRecord

Describes the domain validation records of an Amazon Lightsail SSL/TLS certificate.

See: newDomainValidationRecord smart constructor.

Constructors

DomainValidationRecord'

Fields

resourceRecord :: Maybe ResourceRecord

An object that describes the DNS records to add to your domain's DNS to validate it for the certificate.

domainName :: Maybe Text

The domain name of the certificate validation record. For example, example.com or www.example.com.

newDomainValidationRecord :: DomainValidationRecord

Create a value of DomainValidationRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceRecord:DomainValidationRecord', domainValidationRecord_resourceRecord - An object that describes the DNS records to add to your domain's DNS to validate it for the certificate.

$sel:domainName:DomainValidationRecord', domainValidationRecord_domainName - The domain name of the certificate validation record. For example, example.com or www.example.com.

domainValidationRecord_resourceRecord :: Lens' DomainValidationRecord (Maybe ResourceRecord)

An object that describes the DNS records to add to your domain's DNS to validate it for the certificate.

domainValidationRecord_domainName :: Lens' DomainValidationRecord (Maybe Text)

The domain name of the certificate validation record. For example, example.com or www.example.com.