Network.AWS.Lightsail.Types.HeaderObject

Description

Documentation

data HeaderObject

Describes the request headers that a Lightsail distribution bases caching on.

For the headers that you specify, your distribution caches separate versions of the specified content based on the header values in viewer requests. For example, suppose viewer requests for logo.jpg contain a custom product header that has a value of either acme or apex, and you configure your distribution to cache your content based on values in the product header. Your distribution forwards the product header to the origin and caches the response from the origin once for each header value.

See: newHeaderObject smart constructor.

Constructors

HeaderObject'

Fields

headersAllowList :: Maybe [HeaderEnum]

The specific headers to forward to your distribution's origin.

option :: Maybe ForwardValues

The headers that you want your distribution to forward to your origin and base caching on.

You can configure your distribution to do one of the following:

all - Forward all headers to your origin.

none - Forward only the default headers.

allow-list - Forward only the headers you specify using the headersAllowList parameter.

newHeaderObject :: HeaderObject

Create a value of HeaderObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:headersAllowList:HeaderObject', headerObject_headersAllowList - The specific headers to forward to your distribution's origin.

$sel:option:HeaderObject', headerObject_option - The headers that you want your distribution to forward to your origin and base caching on.

You can configure your distribution to do one of the following:

all - Forward all headers to your origin.

none - Forward only the default headers.

allow-list - Forward only the headers you specify using the headersAllowList parameter.

headerObject_headersAllowList :: Lens' HeaderObject (Maybe [HeaderEnum])

The specific headers to forward to your distribution's origin.

headerObject_option :: Lens' HeaderObject (Maybe ForwardValues)

The headers that you want your distribution to forward to your origin and base caching on.

You can configure your distribution to do one of the following:

all - Forward all headers to your origin.

none - Forward only the default headers.

allow-list - Forward only the headers you specify using the headersAllowList parameter.