Network.AWS.Lightsail.Types.HostKeyAttributes

Description

Documentation

data HostKeyAttributes

Describes the public SSH host keys or the RDP certificate.

See: newHostKeyAttributes smart constructor.

Constructors

HostKeyAttributes'

Fields

algorithm :: Maybe Text

The SSH host key algorithm or the RDP certificate format.

For SSH host keys, the algorithm may be ssh-rsa, ecdsa-sha2-nistp256, ssh-ed25519, etc. For RDP certificates, the algorithm is always x509-cert.

publicKey :: Maybe Text

The public SSH host key or the RDP certificate.

fingerprintSHA256 :: Maybe Text

The SHA-256 fingerprint of the returned SSH host key or RDP certificate.

Example of an SHA-256 SSH fingerprint:

SHA256:KTsMnRBh1IhD17HpdfsbzeGA4jOijm5tyXsMjKVbB8o

Example of an SHA-256 RDP fingerprint:

03:9b:36:9f:4b:de:4e:61:70:fc:7c:c9:78:e7:d2:1a:1c:25:a8:0c:91:f6:7c:e4:d6:a0:85:c8:b4:53:99:68

notValidBefore :: Maybe POSIX

The returned RDP certificate is valid after this point in time.

This value is listed only for RDP certificates.

notValidAfter :: Maybe POSIX

The returned RDP certificate is not valid after this point in time.

This value is listed only for RDP certificates.

fingerprintSHA1 :: Maybe Text

The SHA-1 fingerprint of the returned SSH host key or RDP certificate.

Example of an SHA-1 SSH fingerprint:

SHA1:1CHH6FaAaXjtFOsR/t83vf91SR0

Example of an SHA-1 RDP fingerprint:

af:34:51:fe:09:f0:e0:da:b8:4e:56:ca:60:c2:10:ff:38:06:db:45

witnessedAt :: Maybe POSIX

The time that the SSH host key or RDP certificate was recorded by Lightsail.

newHostKeyAttributes :: HostKeyAttributes

Create a value of HostKeyAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:algorithm:HostKeyAttributes', hostKeyAttributes_algorithm - The SSH host key algorithm or the RDP certificate format.

For SSH host keys, the algorithm may be ssh-rsa, ecdsa-sha2-nistp256, ssh-ed25519, etc. For RDP certificates, the algorithm is always x509-cert.

$sel:publicKey:HostKeyAttributes', hostKeyAttributes_publicKey - The public SSH host key or the RDP certificate.

$sel:fingerprintSHA256:HostKeyAttributes', hostKeyAttributes_fingerprintSHA256 - The SHA-256 fingerprint of the returned SSH host key or RDP certificate.

Example of an SHA-256 SSH fingerprint:

SHA256:KTsMnRBh1IhD17HpdfsbzeGA4jOijm5tyXsMjKVbB8o

Example of an SHA-256 RDP fingerprint:

03:9b:36:9f:4b:de:4e:61:70:fc:7c:c9:78:e7:d2:1a:1c:25:a8:0c:91:f6:7c:e4:d6:a0:85:c8:b4:53:99:68

$sel:notValidBefore:HostKeyAttributes', hostKeyAttributes_notValidBefore - The returned RDP certificate is valid after this point in time.

This value is listed only for RDP certificates.

$sel:notValidAfter:HostKeyAttributes', hostKeyAttributes_notValidAfter - The returned RDP certificate is not valid after this point in time.

This value is listed only for RDP certificates.

$sel:fingerprintSHA1:HostKeyAttributes', hostKeyAttributes_fingerprintSHA1 - The SHA-1 fingerprint of the returned SSH host key or RDP certificate.

Example of an SHA-1 SSH fingerprint:

SHA1:1CHH6FaAaXjtFOsR/t83vf91SR0

Example of an SHA-1 RDP fingerprint:

af:34:51:fe:09:f0:e0:da:b8:4e:56:ca:60:c2:10:ff:38:06:db:45

$sel:witnessedAt:HostKeyAttributes', hostKeyAttributes_witnessedAt - The time that the SSH host key or RDP certificate was recorded by Lightsail.

hostKeyAttributes_algorithm :: Lens' HostKeyAttributes (Maybe Text)

The SSH host key algorithm or the RDP certificate format.

For SSH host keys, the algorithm may be ssh-rsa, ecdsa-sha2-nistp256, ssh-ed25519, etc. For RDP certificates, the algorithm is always x509-cert.

hostKeyAttributes_publicKey :: Lens' HostKeyAttributes (Maybe Text)

The public SSH host key or the RDP certificate.

hostKeyAttributes_fingerprintSHA256 :: Lens' HostKeyAttributes (Maybe Text)

The SHA-256 fingerprint of the returned SSH host key or RDP certificate.

Example of an SHA-256 SSH fingerprint:

SHA256:KTsMnRBh1IhD17HpdfsbzeGA4jOijm5tyXsMjKVbB8o

Example of an SHA-256 RDP fingerprint:

03:9b:36:9f:4b:de:4e:61:70:fc:7c:c9:78:e7:d2:1a:1c:25:a8:0c:91:f6:7c:e4:d6:a0:85:c8:b4:53:99:68

hostKeyAttributes_notValidBefore :: Lens' HostKeyAttributes (Maybe UTCTime)

The returned RDP certificate is valid after this point in time.

This value is listed only for RDP certificates.

hostKeyAttributes_notValidAfter :: Lens' HostKeyAttributes (Maybe UTCTime)

The returned RDP certificate is not valid after this point in time.

This value is listed only for RDP certificates.

hostKeyAttributes_fingerprintSHA1 :: Lens' HostKeyAttributes (Maybe Text)

The SHA-1 fingerprint of the returned SSH host key or RDP certificate.

Example of an SHA-1 SSH fingerprint:

SHA1:1CHH6FaAaXjtFOsR/t83vf91SR0

Example of an SHA-1 RDP fingerprint:

af:34:51:fe:09:f0:e0:da:b8:4e:56:ca:60:c2:10:ff:38:06:db:45

hostKeyAttributes_witnessedAt :: Lens' HostKeyAttributes (Maybe UTCTime)

The time that the SSH host key or RDP certificate was recorded by Lightsail.