Network.AWS.Lightsail.Types.InputOrigin

Description

Documentation

data InputOrigin

Describes the origin resource of an Amazon Lightsail content delivery network (CDN) distribution.

An origin can be a Lightsail instance or load balancer. A distribution pulls content from an origin, caches it, and serves it to viewers via a worldwide network of edge servers.

See: newInputOrigin smart constructor.

Constructors

InputOrigin'

Fields

regionName :: Maybe RegionName

The AWS Region name of the origin resource.

name :: Maybe Text

The name of the origin resource.

protocolPolicy :: Maybe OriginProtocolPolicyEnum

The protocol that your Amazon Lightsail distribution uses when establishing a connection with your origin to pull content.

newInputOrigin :: InputOrigin

Create a value of InputOrigin with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:regionName:InputOrigin', inputOrigin_regionName - The AWS Region name of the origin resource.

$sel:name:InputOrigin', inputOrigin_name - The name of the origin resource.

$sel:protocolPolicy:InputOrigin', inputOrigin_protocolPolicy - The protocol that your Amazon Lightsail distribution uses when establishing a connection with your origin to pull content.

inputOrigin_regionName :: Lens' InputOrigin (Maybe RegionName)

The AWS Region name of the origin resource.

inputOrigin_name :: Lens' InputOrigin (Maybe Text)

The name of the origin resource.

inputOrigin_protocolPolicy :: Lens' InputOrigin (Maybe OriginProtocolPolicyEnum)

The protocol that your Amazon Lightsail distribution uses when establishing a connection with your origin to pull content.