Network.AWS.Lightsail.Types.InstanceAccessDetails

Description

Documentation

data InstanceAccessDetails

The parameters for gaining temporary access to one of your Amazon Lightsail instances.

See: newInstanceAccessDetails smart constructor.

Constructors

InstanceAccessDetails'

Fields

hostKeys :: Maybe [HostKeyAttributes]

Describes the public SSH host keys or the RDP certificate.

instanceName :: Maybe Text

The name of this Amazon Lightsail instance.

expiresAt :: Maybe POSIX

For SSH access, the date on which the temporary keys expire.

privateKey :: Maybe Text

For SSH access, the temporary private key. For OpenSSH clients (e.g., command line SSH), you should save this value to tempkey).

password :: Maybe Text

For RDP access, the password for your Amazon Lightsail instance. Password will be an empty string if the password for your new instance is not ready yet. When you create an instance, it can take up to 15 minutes for the instance to be ready.

If you create an instance using any key pair other than the default (LightsailDefaultKeyPair), password will always be an empty string.

If you change the Administrator password on the instance, Lightsail will continue to return the original password value. When accessing the instance using RDP, you need to manually enter the Administrator password after changing it from the default.

passwordData :: Maybe PasswordData

For a Windows Server-based instance, an object with the data you can use to retrieve your password. This is only needed if password is empty and the instance is not new (and therefore the password is not ready yet). When you create an instance, it can take up to 15 minutes for the instance to be ready.

ipAddress :: Maybe Text

The public IP address of the Amazon Lightsail instance.

certKey :: Maybe Text

For SSH access, the public key to use when accessing your instance For OpenSSH clients (e.g., command line SSH), you should save this value to tempkey-cert.pub.

username :: Maybe Text

The user name to use when logging in to the Amazon Lightsail instance.

protocol :: Maybe InstanceAccessProtocol

The protocol for these Amazon Lightsail instance access details.

newInstanceAccessDetails :: InstanceAccessDetails

Create a value of InstanceAccessDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hostKeys:InstanceAccessDetails', instanceAccessDetails_hostKeys - Describes the public SSH host keys or the RDP certificate.

$sel:instanceName:InstanceAccessDetails', instanceAccessDetails_instanceName - The name of this Amazon Lightsail instance.

$sel:expiresAt:InstanceAccessDetails', instanceAccessDetails_expiresAt - For SSH access, the date on which the temporary keys expire.

$sel:privateKey:InstanceAccessDetails', instanceAccessDetails_privateKey - For SSH access, the temporary private key. For OpenSSH clients (e.g., command line SSH), you should save this value to tempkey).

$sel:password:InstanceAccessDetails', instanceAccessDetails_password - For RDP access, the password for your Amazon Lightsail instance. Password will be an empty string if the password for your new instance is not ready yet. When you create an instance, it can take up to 15 minutes for the instance to be ready.

If you create an instance using any key pair other than the default (LightsailDefaultKeyPair), password will always be an empty string.

If you change the Administrator password on the instance, Lightsail will continue to return the original password value. When accessing the instance using RDP, you need to manually enter the Administrator password after changing it from the default.

$sel:passwordData:InstanceAccessDetails', instanceAccessDetails_passwordData - For a Windows Server-based instance, an object with the data you can use to retrieve your password. This is only needed if password is empty and the instance is not new (and therefore the password is not ready yet). When you create an instance, it can take up to 15 minutes for the instance to be ready.

$sel:ipAddress:InstanceAccessDetails', instanceAccessDetails_ipAddress - The public IP address of the Amazon Lightsail instance.

$sel:certKey:InstanceAccessDetails', instanceAccessDetails_certKey - For SSH access, the public key to use when accessing your instance For OpenSSH clients (e.g., command line SSH), you should save this value to tempkey-cert.pub.

$sel:username:InstanceAccessDetails', instanceAccessDetails_username - The user name to use when logging in to the Amazon Lightsail instance.

$sel:protocol:InstanceAccessDetails', instanceAccessDetails_protocol - The protocol for these Amazon Lightsail instance access details.

instanceAccessDetails_hostKeys :: Lens' InstanceAccessDetails (Maybe [HostKeyAttributes])

Describes the public SSH host keys or the RDP certificate.

instanceAccessDetails_instanceName :: Lens' InstanceAccessDetails (Maybe Text)

The name of this Amazon Lightsail instance.

instanceAccessDetails_expiresAt :: Lens' InstanceAccessDetails (Maybe UTCTime)

For SSH access, the date on which the temporary keys expire.

instanceAccessDetails_privateKey :: Lens' InstanceAccessDetails (Maybe Text)

For SSH access, the temporary private key. For OpenSSH clients (e.g., command line SSH), you should save this value to tempkey).

instanceAccessDetails_password :: Lens' InstanceAccessDetails (Maybe Text)

For RDP access, the password for your Amazon Lightsail instance. Password will be an empty string if the password for your new instance is not ready yet. When you create an instance, it can take up to 15 minutes for the instance to be ready.

If you create an instance using any key pair other than the default (LightsailDefaultKeyPair), password will always be an empty string.

If you change the Administrator password on the instance, Lightsail will continue to return the original password value. When accessing the instance using RDP, you need to manually enter the Administrator password after changing it from the default.

instanceAccessDetails_passwordData :: Lens' InstanceAccessDetails (Maybe PasswordData)

For a Windows Server-based instance, an object with the data you can use to retrieve your password. This is only needed if password is empty and the instance is not new (and therefore the password is not ready yet). When you create an instance, it can take up to 15 minutes for the instance to be ready.

instanceAccessDetails_ipAddress :: Lens' InstanceAccessDetails (Maybe Text)

The public IP address of the Amazon Lightsail instance.

instanceAccessDetails_certKey :: Lens' InstanceAccessDetails (Maybe Text)

For SSH access, the public key to use when accessing your instance For OpenSSH clients (e.g., command line SSH), you should save this value to tempkey-cert.pub.

instanceAccessDetails_username :: Lens' InstanceAccessDetails (Maybe Text)

The user name to use when logging in to the Amazon Lightsail instance.

instanceAccessDetails_protocol :: Lens' InstanceAccessDetails (Maybe InstanceAccessProtocol)

The protocol for these Amazon Lightsail instance access details.