Network.AWS.Lightsail.Types.InstanceEntry

Description

Documentation

data InstanceEntry

Describes the Amazon Elastic Compute Cloud instance and related resources to be created using the create cloud formation stack operation.

See: newInstanceEntry smart constructor.

Constructors

InstanceEntry'

Fields

userData :: Maybe Text

A launch script you can create that configures a server with additional user data. For example, you might want to run apt-get -y update.

Depending on the machine image you choose, the command to get software on your instance varies. Amazon Linux and CentOS use yum, Debian and Ubuntu use apt-get, and FreeBSD uses pkg.

sourceName :: Text

The name of the export snapshot record, which contains the exported Lightsail instance snapshot that will be used as the source of the new Amazon EC2 instance.

Use the get export snapshot records operation to get a list of export snapshot records that you can use to create a CloudFormation stack.

instanceType :: Text

The instance type (e.g., t2.micro) to use for the new Amazon EC2 instance.

portInfoSource :: PortInfoSourceType

The port configuration to use for the new Amazon EC2 instance.

The following configuration options are available:

DEFAULT - Use the default firewall settings from the Lightsail instance blueprint. If this is specified, then IPv4 and IPv6 will be configured for the new instance that is created in Amazon EC2.

INSTANCE - Use the configured firewall settings from the source Lightsail instance. If this is specified, the new instance that is created in Amazon EC2 will be configured to match the configuration of the source Lightsail instance. For example, if the source instance is configured for dual-stack (IPv4 and IPv6), then IPv4 and IPv6 will be configured for the new instance that is created in Amazon EC2. If the source instance is configured for IPv4 only, then only IPv4 will be configured for the new instance that is created in Amazon EC2.

NONE - Use the default Amazon EC2 security group. If this is specified, then only IPv4 will be configured for the new instance that is created in Amazon EC2.

CLOSED - All ports closed. If this is specified, then only IPv4 will be configured for the new instance that is created in Amazon EC2.

If you configured lightsail-connect as a cidrListAliases on your instance, or if you chose to allow the Lightsail browser-based SSH or RDP clients to connect to your instance, that configuration is not carried over to your new Amazon EC2 instance.

availabilityZone :: Text

The Availability Zone for the new Amazon EC2 instance.

newInstanceEntry

Arguments

:: Text
$sel:sourceName:InstanceEntry'

-> Text
$sel:instanceType:InstanceEntry'

-> PortInfoSourceType
$sel:portInfoSource:InstanceEntry'

-> Text
$sel:availabilityZone:InstanceEntry'

-> InstanceEntry

Create a value of InstanceEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userData:InstanceEntry', instanceEntry_userData - A launch script you can create that configures a server with additional user data. For example, you might want to run apt-get -y update.

Depending on the machine image you choose, the command to get software on your instance varies. Amazon Linux and CentOS use yum, Debian and Ubuntu use apt-get, and FreeBSD uses pkg.

$sel:sourceName:InstanceEntry', instanceEntry_sourceName - The name of the export snapshot record, which contains the exported Lightsail instance snapshot that will be used as the source of the new Amazon EC2 instance.

Use the get export snapshot records operation to get a list of export snapshot records that you can use to create a CloudFormation stack.

$sel:instanceType:InstanceEntry', instanceEntry_instanceType - The instance type (e.g., t2.micro) to use for the new Amazon EC2 instance.

$sel:portInfoSource:InstanceEntry', instanceEntry_portInfoSource - The port configuration to use for the new Amazon EC2 instance.

The following configuration options are available:

DEFAULT - Use the default firewall settings from the Lightsail instance blueprint. If this is specified, then IPv4 and IPv6 will be configured for the new instance that is created in Amazon EC2.

INSTANCE - Use the configured firewall settings from the source Lightsail instance. If this is specified, the new instance that is created in Amazon EC2 will be configured to match the configuration of the source Lightsail instance. For example, if the source instance is configured for dual-stack (IPv4 and IPv6), then IPv4 and IPv6 will be configured for the new instance that is created in Amazon EC2. If the source instance is configured for IPv4 only, then only IPv4 will be configured for the new instance that is created in Amazon EC2.

NONE - Use the default Amazon EC2 security group. If this is specified, then only IPv4 will be configured for the new instance that is created in Amazon EC2.

CLOSED - All ports closed. If this is specified, then only IPv4 will be configured for the new instance that is created in Amazon EC2.

If you configured lightsail-connect as a cidrListAliases on your instance, or if you chose to allow the Lightsail browser-based SSH or RDP clients to connect to your instance, that configuration is not carried over to your new Amazon EC2 instance.

$sel:availabilityZone:InstanceEntry', instanceEntry_availabilityZone - The Availability Zone for the new Amazon EC2 instance.

instanceEntry_userData :: Lens' InstanceEntry (Maybe Text)

A launch script you can create that configures a server with additional user data. For example, you might want to run apt-get -y update.

Depending on the machine image you choose, the command to get software on your instance varies. Amazon Linux and CentOS use yum, Debian and Ubuntu use apt-get, and FreeBSD uses pkg.

instanceEntry_sourceName :: Lens' InstanceEntry Text

The name of the export snapshot record, which contains the exported Lightsail instance snapshot that will be used as the source of the new Amazon EC2 instance.

Use the get export snapshot records operation to get a list of export snapshot records that you can use to create a CloudFormation stack.

instanceEntry_instanceType :: Lens' InstanceEntry Text

The instance type (e.g., t2.micro) to use for the new Amazon EC2 instance.

instanceEntry_portInfoSource :: Lens' InstanceEntry PortInfoSourceType

The port configuration to use for the new Amazon EC2 instance.

The following configuration options are available:

DEFAULT - Use the default firewall settings from the Lightsail instance blueprint. If this is specified, then IPv4 and IPv6 will be configured for the new instance that is created in Amazon EC2.

INSTANCE - Use the configured firewall settings from the source Lightsail instance. If this is specified, the new instance that is created in Amazon EC2 will be configured to match the configuration of the source Lightsail instance. For example, if the source instance is configured for dual-stack (IPv4 and IPv6), then IPv4 and IPv6 will be configured for the new instance that is created in Amazon EC2. If the source instance is configured for IPv4 only, then only IPv4 will be configured for the new instance that is created in Amazon EC2.

NONE - Use the default Amazon EC2 security group. If this is specified, then only IPv4 will be configured for the new instance that is created in Amazon EC2.

CLOSED - All ports closed. If this is specified, then only IPv4 will be configured for the new instance that is created in Amazon EC2.

If you configured lightsail-connect as a cidrListAliases on your instance, or if you chose to allow the Lightsail browser-based SSH or RDP clients to connect to your instance, that configuration is not carried over to your new Amazon EC2 instance.

instanceEntry_availabilityZone :: Lens' InstanceEntry Text

The Availability Zone for the new Amazon EC2 instance.