Network.AWS.Lightsail.Types.InstancePortInfo

Description

Documentation

data InstancePortInfo

Describes information about ports for an Amazon Lightsail instance.

See: newInstancePortInfo smart constructor.

Constructors

InstancePortInfo'

Fields

fromPort :: Maybe Int

The first port in a range of open ports on an instance.

Allowed ports:

TCP and UDP - 0 to 65535

ICMP - The ICMP type for IPv4 addresses. For example, specify 8 as the fromPort (ICMP type), and -1 as the toPort (ICMP code), to enable ICMP Ping. For more information, see Control Messages on Wikipedia.

ICMPv6 - The ICMP type for IPv6 addresses. For example, specify 128 as the fromPort (ICMPv6 type), and 0 as toPort (ICMPv6 code). For more information, see Internet Control Message Protocol for IPv6.

cidrs :: Maybe [Text]

The IPv4 address, or range of IPv4 addresses (in CIDR notation) that are allowed to connect to an instance through the ports, and the protocol.

The ipv6Cidrs parameter lists the IPv6 addresses that are allowed to connect to an instance.

For more information about CIDR block notation, see Classless Inter-Domain Routing on Wikipedia.

commonName :: Maybe Text

The common name of the port information.

protocol :: Maybe NetworkProtocol

The IP protocol name.

The name can be one of the following:

tcp - Transmission Control Protocol (TCP) provides reliable, ordered, and error-checked delivery of streamed data between applications running on hosts communicating by an IP network. If you have an application that doesn't require reliable data stream service, use UDP instead.

all - All transport layer protocol types. For more general information, see Transport layer on Wikipedia.

udp - With User Datagram Protocol (UDP), computer applications can send messages (or datagrams) to other hosts on an Internet Protocol (IP) network. Prior communications are not required to set up transmission channels or data paths. Applications that don't require reliable data stream service can use UDP, which provides a connectionless datagram service that emphasizes reduced latency over reliability. If you do require reliable data stream service, use TCP instead.

icmp - Internet Control Message Protocol (ICMP) is used to send error messages and operational information indicating success or failure when communicating with an instance. For example, an error is indicated when an instance could not be reached. When you specify icmp as the protocol, you must specify the ICMP type using the fromPort parameter, and ICMP code using the toPort parameter.

cidrListAliases :: Maybe [Text]

An alias that defines access for a preconfigured range of IP addresses.

The only alias currently supported is lightsail-connect, which allows IP addresses of the browser-based RDP/SSH client in the Lightsail console to connect to your instance.

ipv6Cidrs :: Maybe [Text]

The IPv6 address, or range of IPv6 addresses (in CIDR notation) that are allowed to connect to an instance through the ports, and the protocol. Only devices with an IPv6 address can connect to an instance through IPv6; otherwise, IPv4 should be used.

The cidrs parameter lists the IPv4 addresses that are allowed to connect to an instance.

For more information about CIDR block notation, see Classless Inter-Domain Routing on Wikipedia.

accessDirection :: Maybe AccessDirection

The access direction (inbound or outbound).

Lightsail currently supports only inbound access direction.

accessType :: Maybe PortAccessType

The type of access (Public or Private).

toPort :: Maybe Int

The last port in a range of open ports on an instance.

Allowed ports:

TCP and UDP - 0 to 65535

ICMP - The ICMP code for IPv4 addresses. For example, specify 8 as the fromPort (ICMP type), and -1 as the toPort (ICMP code), to enable ICMP Ping. For more information, see Control Messages on Wikipedia.

ICMPv6 - The ICMP code for IPv6 addresses. For example, specify 128 as the fromPort (ICMPv6 type), and 0 as toPort (ICMPv6 code). For more information, see Internet Control Message Protocol for IPv6.

accessFrom :: Maybe Text

The location from which access is allowed. For example, Anywhere (0.0.0.0/0), or Custom if a specific IP address or range of IP addresses is allowed.

newInstancePortInfo :: InstancePortInfo

Create a value of InstancePortInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fromPort:InstancePortInfo', instancePortInfo_fromPort - The first port in a range of open ports on an instance.

Allowed ports:

TCP and UDP - 0 to 65535

ICMP - The ICMP type for IPv4 addresses. For example, specify 8 as the fromPort (ICMP type), and -1 as the toPort (ICMP code), to enable ICMP Ping. For more information, see Control Messages on Wikipedia.

ICMPv6 - The ICMP type for IPv6 addresses. For example, specify 128 as the fromPort (ICMPv6 type), and 0 as toPort (ICMPv6 code). For more information, see Internet Control Message Protocol for IPv6.

$sel:cidrs:InstancePortInfo', instancePortInfo_cidrs - The IPv4 address, or range of IPv4 addresses (in CIDR notation) that are allowed to connect to an instance through the ports, and the protocol.

The ipv6Cidrs parameter lists the IPv6 addresses that are allowed to connect to an instance.

For more information about CIDR block notation, see Classless Inter-Domain Routing on Wikipedia.

$sel:commonName:InstancePortInfo', instancePortInfo_commonName - The common name of the port information.

$sel:protocol:InstancePortInfo', instancePortInfo_protocol - The IP protocol name.

The name can be one of the following:

tcp - Transmission Control Protocol (TCP) provides reliable, ordered, and error-checked delivery of streamed data between applications running on hosts communicating by an IP network. If you have an application that doesn't require reliable data stream service, use UDP instead.

all - All transport layer protocol types. For more general information, see Transport layer on Wikipedia.

udp - With User Datagram Protocol (UDP), computer applications can send messages (or datagrams) to other hosts on an Internet Protocol (IP) network. Prior communications are not required to set up transmission channels or data paths. Applications that don't require reliable data stream service can use UDP, which provides a connectionless datagram service that emphasizes reduced latency over reliability. If you do require reliable data stream service, use TCP instead.

icmp - Internet Control Message Protocol (ICMP) is used to send error messages and operational information indicating success or failure when communicating with an instance. For example, an error is indicated when an instance could not be reached. When you specify icmp as the protocol, you must specify the ICMP type using the fromPort parameter, and ICMP code using the toPort parameter.

$sel:cidrListAliases:InstancePortInfo', instancePortInfo_cidrListAliases - An alias that defines access for a preconfigured range of IP addresses.

The only alias currently supported is lightsail-connect, which allows IP addresses of the browser-based RDP/SSH client in the Lightsail console to connect to your instance.

$sel:ipv6Cidrs:InstancePortInfo', instancePortInfo_ipv6Cidrs - The IPv6 address, or range of IPv6 addresses (in CIDR notation) that are allowed to connect to an instance through the ports, and the protocol. Only devices with an IPv6 address can connect to an instance through IPv6; otherwise, IPv4 should be used.

The cidrs parameter lists the IPv4 addresses that are allowed to connect to an instance.

For more information about CIDR block notation, see Classless Inter-Domain Routing on Wikipedia.

$sel:accessDirection:InstancePortInfo', instancePortInfo_accessDirection - The access direction (inbound or outbound).

Lightsail currently supports only inbound access direction.

$sel:accessType:InstancePortInfo', instancePortInfo_accessType - The type of access (Public or Private).

$sel:toPort:InstancePortInfo', instancePortInfo_toPort - The last port in a range of open ports on an instance.

Allowed ports:

TCP and UDP - 0 to 65535

ICMP - The ICMP code for IPv4 addresses. For example, specify 8 as the fromPort (ICMP type), and -1 as the toPort (ICMP code), to enable ICMP Ping. For more information, see Control Messages on Wikipedia.

ICMPv6 - The ICMP code for IPv6 addresses. For example, specify 128 as the fromPort (ICMPv6 type), and 0 as toPort (ICMPv6 code). For more information, see Internet Control Message Protocol for IPv6.

$sel:accessFrom:InstancePortInfo', instancePortInfo_accessFrom - The location from which access is allowed. For example, Anywhere (0.0.0.0/0), or Custom if a specific IP address or range of IP addresses is allowed.

instancePortInfo_fromPort :: Lens' InstancePortInfo (Maybe Int)

The first port in a range of open ports on an instance.

Allowed ports:

TCP and UDP - 0 to 65535

ICMP - The ICMP type for IPv4 addresses. For example, specify 8 as the fromPort (ICMP type), and -1 as the toPort (ICMP code), to enable ICMP Ping. For more information, see Control Messages on Wikipedia.

ICMPv6 - The ICMP type for IPv6 addresses. For example, specify 128 as the fromPort (ICMPv6 type), and 0 as toPort (ICMPv6 code). For more information, see Internet Control Message Protocol for IPv6.

instancePortInfo_cidrs :: Lens' InstancePortInfo (Maybe [Text])

The IPv4 address, or range of IPv4 addresses (in CIDR notation) that are allowed to connect to an instance through the ports, and the protocol.

The ipv6Cidrs parameter lists the IPv6 addresses that are allowed to connect to an instance.

For more information about CIDR block notation, see Classless Inter-Domain Routing on Wikipedia.

instancePortInfo_commonName :: Lens' InstancePortInfo (Maybe Text)

The common name of the port information.

instancePortInfo_protocol :: Lens' InstancePortInfo (Maybe NetworkProtocol)

The IP protocol name.

The name can be one of the following:

tcp - Transmission Control Protocol (TCP) provides reliable, ordered, and error-checked delivery of streamed data between applications running on hosts communicating by an IP network. If you have an application that doesn't require reliable data stream service, use UDP instead.

all - All transport layer protocol types. For more general information, see Transport layer on Wikipedia.

udp - With User Datagram Protocol (UDP), computer applications can send messages (or datagrams) to other hosts on an Internet Protocol (IP) network. Prior communications are not required to set up transmission channels or data paths. Applications that don't require reliable data stream service can use UDP, which provides a connectionless datagram service that emphasizes reduced latency over reliability. If you do require reliable data stream service, use TCP instead.

icmp - Internet Control Message Protocol (ICMP) is used to send error messages and operational information indicating success or failure when communicating with an instance. For example, an error is indicated when an instance could not be reached. When you specify icmp as the protocol, you must specify the ICMP type using the fromPort parameter, and ICMP code using the toPort parameter.

instancePortInfo_cidrListAliases :: Lens' InstancePortInfo (Maybe [Text])

An alias that defines access for a preconfigured range of IP addresses.

The only alias currently supported is lightsail-connect, which allows IP addresses of the browser-based RDP/SSH client in the Lightsail console to connect to your instance.

instancePortInfo_ipv6Cidrs :: Lens' InstancePortInfo (Maybe [Text])

The IPv6 address, or range of IPv6 addresses (in CIDR notation) that are allowed to connect to an instance through the ports, and the protocol. Only devices with an IPv6 address can connect to an instance through IPv6; otherwise, IPv4 should be used.

The cidrs parameter lists the IPv4 addresses that are allowed to connect to an instance.

For more information about CIDR block notation, see Classless Inter-Domain Routing on Wikipedia.

instancePortInfo_accessDirection :: Lens' InstancePortInfo (Maybe AccessDirection)

The access direction (inbound or outbound).

Lightsail currently supports only inbound access direction.

instancePortInfo_accessType :: Lens' InstancePortInfo (Maybe PortAccessType)

The type of access (Public or Private).

instancePortInfo_toPort :: Lens' InstancePortInfo (Maybe Int)

The last port in a range of open ports on an instance.

Allowed ports:

TCP and UDP - 0 to 65535

ICMP - The ICMP code for IPv4 addresses. For example, specify 8 as the fromPort (ICMP type), and -1 as the toPort (ICMP code), to enable ICMP Ping. For more information, see Control Messages on Wikipedia.

ICMPv6 - The ICMP code for IPv6 addresses. For example, specify 128 as the fromPort (ICMPv6 type), and 0 as toPort (ICMPv6 code). For more information, see Internet Control Message Protocol for IPv6.

instancePortInfo_accessFrom :: Lens' InstancePortInfo (Maybe Text)

The location from which access is allowed. For example, Anywhere (0.0.0.0/0), or Custom if a specific IP address or range of IP addresses is allowed.