Network.AWS.Lightsail.Types.InstanceSnapshotInfo

Description

Documentation

data InstanceSnapshotInfo

Describes an instance snapshot.

See: newInstanceSnapshotInfo smart constructor.

Constructors

InstanceSnapshotInfo'

Fields

fromDiskInfo :: Maybe [DiskInfo]

A list of objects describing the disks that were attached to the source instance.

fromBundleId :: Maybe Text

The bundle ID from which the source instance was created (e.g., micro_1_0).

fromBlueprintId :: Maybe Text

The blueprint ID from which the source instance (e.g., os_debian_8_3).

newInstanceSnapshotInfo :: InstanceSnapshotInfo

Create a value of InstanceSnapshotInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fromDiskInfo:InstanceSnapshotInfo', instanceSnapshotInfo_fromDiskInfo - A list of objects describing the disks that were attached to the source instance.

$sel:fromBundleId:InstanceSnapshotInfo', instanceSnapshotInfo_fromBundleId - The bundle ID from which the source instance was created (e.g., micro_1_0).

$sel:fromBlueprintId:InstanceSnapshotInfo', instanceSnapshotInfo_fromBlueprintId - The blueprint ID from which the source instance (e.g., os_debian_8_3).

instanceSnapshotInfo_fromDiskInfo :: Lens' InstanceSnapshotInfo (Maybe [DiskInfo])

A list of objects describing the disks that were attached to the source instance.

instanceSnapshotInfo_fromBundleId :: Lens' InstanceSnapshotInfo (Maybe Text)

The bundle ID from which the source instance was created (e.g., micro_1_0).

instanceSnapshotInfo_fromBlueprintId :: Lens' InstanceSnapshotInfo (Maybe Text)

The blueprint ID from which the source instance (e.g., os_debian_8_3).