Network.AWS.Lightsail.Types.KeyPair

Description

Documentation

data KeyPair

Describes an SSH key pair.

See: newKeyPair smart constructor.

Constructors

KeyPair'

Fields

resourceType :: Maybe ResourceType

The resource type (usually KeyPair).

arn :: Maybe Text

The Amazon Resource Name (ARN) of the key pair (e.g., arn:aws:lightsail:us-east-2:123456789101:KeyPair/05859e3d-331d-48ba-9034-12345EXAMPLE).

createdAt :: Maybe POSIX

The timestamp when the key pair was created (e.g., 1479816991.349).

location :: Maybe ResourceLocation

The region name and Availability Zone where the key pair was created.

fingerprint :: Maybe Text

The RSA fingerprint of the key pair.

name :: Maybe Text

The friendly name of the SSH key pair.

supportCode :: Maybe Text

The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

tags :: Maybe [Tag]

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

newKeyPair :: KeyPair

Create a value of KeyPair with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:KeyPair', keyPair_resourceType - The resource type (usually KeyPair).

$sel:arn:KeyPair', keyPair_arn - The Amazon Resource Name (ARN) of the key pair (e.g., arn:aws:lightsail:us-east-2:123456789101:KeyPair/05859e3d-331d-48ba-9034-12345EXAMPLE).

$sel:createdAt:KeyPair', keyPair_createdAt - The timestamp when the key pair was created (e.g., 1479816991.349).

$sel:location:KeyPair', keyPair_location - The region name and Availability Zone where the key pair was created.

$sel:fingerprint:KeyPair', keyPair_fingerprint - The RSA fingerprint of the key pair.

$sel:name:KeyPair', keyPair_name - The friendly name of the SSH key pair.

$sel:supportCode:KeyPair', keyPair_supportCode - The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

$sel:tags:KeyPair', keyPair_tags - The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

keyPair_resourceType :: Lens' KeyPair (Maybe ResourceType)

The resource type (usually KeyPair).

keyPair_arn :: Lens' KeyPair (Maybe Text)

The Amazon Resource Name (ARN) of the key pair (e.g., arn:aws:lightsail:us-east-2:123456789101:KeyPair/05859e3d-331d-48ba-9034-12345EXAMPLE).

keyPair_createdAt :: Lens' KeyPair (Maybe UTCTime)

The timestamp when the key pair was created (e.g., 1479816991.349).

keyPair_location :: Lens' KeyPair (Maybe ResourceLocation)

The region name and Availability Zone where the key pair was created.

keyPair_fingerprint :: Lens' KeyPair (Maybe Text)

The RSA fingerprint of the key pair.

keyPair_name :: Lens' KeyPair (Maybe Text)

The friendly name of the SSH key pair.

keyPair_supportCode :: Lens' KeyPair (Maybe Text)

The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

keyPair_tags :: Lens' KeyPair (Maybe [Tag])

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.