Network.AWS.Lightsail.Types.LightsailDistribution

Description

Documentation

data LightsailDistribution

Describes an Amazon Lightsail content delivery network (CDN) distribution.

See: newLightsailDistribution smart constructor.

Constructors

LightsailDistribution'

Fields

isEnabled :: Maybe Bool

Indicates whether the distribution is enabled.

ipAddressType :: Maybe IpAddressType

The IP address type of the distribution.

The possible values are ipv4 for IPv4 only, and dualstack for IPv4 and IPv6.

status :: Maybe Text

The status of the distribution.

originPublicDNS :: Maybe Text

The public DNS of the origin.

origin :: Maybe Origin

An object that describes the origin resource of the distribution, such as a Lightsail instance or load balancer.

The distribution pulls, caches, and serves content from the origin.

bundleId :: Maybe Text

The ID of the bundle currently applied to the distribution.

alternativeDomainNames :: Maybe [Text]

The alternate domain names of the distribution.

createdAt :: Maybe POSIX

The timestamp when the distribution was created.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the distribution.

cacheBehaviorSettings :: Maybe CacheSettings

An object that describes the cache behavior settings of the distribution.

supportCode :: Maybe Text

The support code. Include this code in your email to support when you have questions about your Lightsail distribution. This code enables our support team to look up your Lightsail information more easily.

resourceType :: Maybe ResourceType

The Lightsail resource type (e.g., Distribution).

domainName :: Maybe Text

The domain name of the distribution.

name :: Maybe Text

The name of the distribution.

ableToUpdateBundle :: Maybe Bool

Indicates whether the bundle that is currently applied to your distribution, specified using the distributionName parameter, can be changed to another bundle.

Use the UpdateDistributionBundle action to change your distribution's bundle.

tags :: Maybe [Tag]

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

cacheBehaviors :: Maybe [CacheBehaviorPerPath]

An array of objects that describe the per-path cache behavior of the distribution.

defaultCacheBehavior :: Maybe CacheBehavior

An object that describes the default cache behavior of the distribution.

location :: Maybe ResourceLocation

An object that describes the location of the distribution, such as the AWS Region and Availability Zone.

Lightsail distributions are global resources that can reference an origin in any AWS Region, and distribute its content globally. However, all distributions are located in the us-east-1 Region.

certificateName :: Maybe Text

The name of the SSL/TLS certificate attached to the distribution, if any.

newLightsailDistribution :: LightsailDistribution

Create a value of LightsailDistribution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isEnabled:LightsailDistribution', lightsailDistribution_isEnabled - Indicates whether the distribution is enabled.

$sel:ipAddressType:LightsailDistribution', lightsailDistribution_ipAddressType - The IP address type of the distribution.

The possible values are ipv4 for IPv4 only, and dualstack for IPv4 and IPv6.

$sel:status:LightsailDistribution', lightsailDistribution_status - The status of the distribution.

$sel:originPublicDNS:LightsailDistribution', lightsailDistribution_originPublicDNS - The public DNS of the origin.

$sel:origin:LightsailDistribution', lightsailDistribution_origin - An object that describes the origin resource of the distribution, such as a Lightsail instance or load balancer.

The distribution pulls, caches, and serves content from the origin.

$sel:bundleId:LightsailDistribution', lightsailDistribution_bundleId - The ID of the bundle currently applied to the distribution.

$sel:alternativeDomainNames:LightsailDistribution', lightsailDistribution_alternativeDomainNames - The alternate domain names of the distribution.

$sel:createdAt:LightsailDistribution', lightsailDistribution_createdAt - The timestamp when the distribution was created.

$sel:arn:LightsailDistribution', lightsailDistribution_arn - The Amazon Resource Name (ARN) of the distribution.

$sel:cacheBehaviorSettings:LightsailDistribution', lightsailDistribution_cacheBehaviorSettings - An object that describes the cache behavior settings of the distribution.

$sel:supportCode:LightsailDistribution', lightsailDistribution_supportCode - The support code. Include this code in your email to support when you have questions about your Lightsail distribution. This code enables our support team to look up your Lightsail information more easily.

$sel:resourceType:LightsailDistribution', lightsailDistribution_resourceType - The Lightsail resource type (e.g., Distribution).

$sel:domainName:LightsailDistribution', lightsailDistribution_domainName - The domain name of the distribution.

$sel:name:LightsailDistribution', lightsailDistribution_name - The name of the distribution.

$sel:ableToUpdateBundle:LightsailDistribution', lightsailDistribution_ableToUpdateBundle - Indicates whether the bundle that is currently applied to your distribution, specified using the distributionName parameter, can be changed to another bundle.

Use the UpdateDistributionBundle action to change your distribution's bundle.

$sel:tags:LightsailDistribution', lightsailDistribution_tags - The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

$sel:cacheBehaviors:LightsailDistribution', lightsailDistribution_cacheBehaviors - An array of objects that describe the per-path cache behavior of the distribution.

$sel:defaultCacheBehavior:LightsailDistribution', lightsailDistribution_defaultCacheBehavior - An object that describes the default cache behavior of the distribution.

$sel:location:LightsailDistribution', lightsailDistribution_location - An object that describes the location of the distribution, such as the AWS Region and Availability Zone.

Lightsail distributions are global resources that can reference an origin in any AWS Region, and distribute its content globally. However, all distributions are located in the us-east-1 Region.

$sel:certificateName:LightsailDistribution', lightsailDistribution_certificateName - The name of the SSL/TLS certificate attached to the distribution, if any.

lightsailDistribution_isEnabled :: Lens' LightsailDistribution (Maybe Bool)

Indicates whether the distribution is enabled.

lightsailDistribution_ipAddressType :: Lens' LightsailDistribution (Maybe IpAddressType)

The IP address type of the distribution.

The possible values are ipv4 for IPv4 only, and dualstack for IPv4 and IPv6.

lightsailDistribution_status :: Lens' LightsailDistribution (Maybe Text)

The status of the distribution.

lightsailDistribution_originPublicDNS :: Lens' LightsailDistribution (Maybe Text)

The public DNS of the origin.

lightsailDistribution_origin :: Lens' LightsailDistribution (Maybe Origin)

An object that describes the origin resource of the distribution, such as a Lightsail instance or load balancer.

The distribution pulls, caches, and serves content from the origin.

lightsailDistribution_bundleId :: Lens' LightsailDistribution (Maybe Text)

The ID of the bundle currently applied to the distribution.

lightsailDistribution_alternativeDomainNames :: Lens' LightsailDistribution (Maybe [Text])

The alternate domain names of the distribution.

lightsailDistribution_createdAt :: Lens' LightsailDistribution (Maybe UTCTime)

The timestamp when the distribution was created.

lightsailDistribution_arn :: Lens' LightsailDistribution (Maybe Text)

The Amazon Resource Name (ARN) of the distribution.

lightsailDistribution_cacheBehaviorSettings :: Lens' LightsailDistribution (Maybe CacheSettings)

An object that describes the cache behavior settings of the distribution.

lightsailDistribution_supportCode :: Lens' LightsailDistribution (Maybe Text)

The support code. Include this code in your email to support when you have questions about your Lightsail distribution. This code enables our support team to look up your Lightsail information more easily.

lightsailDistribution_resourceType :: Lens' LightsailDistribution (Maybe ResourceType)

The Lightsail resource type (e.g., Distribution).

lightsailDistribution_domainName :: Lens' LightsailDistribution (Maybe Text)

The domain name of the distribution.

lightsailDistribution_name :: Lens' LightsailDistribution (Maybe Text)

The name of the distribution.

lightsailDistribution_ableToUpdateBundle :: Lens' LightsailDistribution (Maybe Bool)

Indicates whether the bundle that is currently applied to your distribution, specified using the distributionName parameter, can be changed to another bundle.

Use the UpdateDistributionBundle action to change your distribution's bundle.

lightsailDistribution_tags :: Lens' LightsailDistribution (Maybe [Tag])

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

lightsailDistribution_cacheBehaviors :: Lens' LightsailDistribution (Maybe [CacheBehaviorPerPath])

An array of objects that describe the per-path cache behavior of the distribution.

lightsailDistribution_defaultCacheBehavior :: Lens' LightsailDistribution (Maybe CacheBehavior)

An object that describes the default cache behavior of the distribution.

lightsailDistribution_location :: Lens' LightsailDistribution (Maybe ResourceLocation)

An object that describes the location of the distribution, such as the AWS Region and Availability Zone.

Lightsail distributions are global resources that can reference an origin in any AWS Region, and distribute its content globally. However, all distributions are located in the us-east-1 Region.

lightsailDistribution_certificateName :: Lens' LightsailDistribution (Maybe Text)

The name of the SSL/TLS certificate attached to the distribution, if any.