Network.AWS.Lightsail.Types.LoadBalancer

Description

Documentation

data LoadBalancer

Describes a load balancer.

See: newLoadBalancer smart constructor.

Constructors

LoadBalancer'

Fields

healthCheckPath :: Maybe Text

The path you specified to perform your health checks. If no path is specified, the load balancer tries to make a request to the default (root) page.

state :: Maybe LoadBalancerState

The status of your load balancer. Valid values are below.

resourceType :: Maybe ResourceType

The resource type (e.g., LoadBalancer.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the load balancer.

createdAt :: Maybe POSIX

The date when your load balancer was created.

location :: Maybe ResourceLocation

The AWS Region where your load balancer was created (e.g., us-east-2a). Lightsail automatically creates your load balancer across Availability Zones.

instancePort :: Maybe Int

The port where the load balancer will direct traffic to your Lightsail instances. For HTTP traffic, it's port 80. For HTTPS traffic, it's port 443.

configurationOptions :: Maybe (HashMap LoadBalancerAttributeName Text)

A string to string map of the configuration options for your load balancer. Valid values are listed below.

protocol :: Maybe LoadBalancerProtocol

The protocol you have enabled for your load balancer. Valid values are below.

You can't just have HTTP_HTTPS, but you can have just HTTP.

tlsCertificateSummaries :: Maybe [LoadBalancerTlsCertificateSummary]

An array of LoadBalancerTlsCertificateSummary objects that provide additional information about the SSL/TLS certificates. For example, if true, the certificate is attached to the load balancer.

name :: Maybe Text

The name of the load balancer (e.g., my-load-balancer).

ipAddressType :: Maybe IpAddressType

The IP address type of the load balancer.

The possible values are ipv4 for IPv4 only, and dualstack for IPv4 and IPv6.

supportCode :: Maybe Text

The support code. Include this code in your email to support when you have questions about your Lightsail load balancer. This code enables our support team to look up your Lightsail information more easily.

publicPorts :: Maybe [Int]

An array of public port settings for your load balancer. For HTTP, use port 80. For HTTPS, use port 443.

dnsName :: Maybe Text

The DNS name of your Lightsail load balancer.

instanceHealthSummary :: Maybe [InstanceHealthSummary]

An array of InstanceHealthSummary objects describing the health of the load balancer.

tags :: Maybe [Tag]

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

newLoadBalancer :: LoadBalancer

Create a value of LoadBalancer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:healthCheckPath:LoadBalancer', loadBalancer_healthCheckPath - The path you specified to perform your health checks. If no path is specified, the load balancer tries to make a request to the default (root) page.

$sel:state:LoadBalancer', loadBalancer_state - The status of your load balancer. Valid values are below.

$sel:resourceType:LoadBalancer', loadBalancer_resourceType - The resource type (e.g., LoadBalancer.

$sel:arn:LoadBalancer', loadBalancer_arn - The Amazon Resource Name (ARN) of the load balancer.

$sel:createdAt:LoadBalancer', loadBalancer_createdAt - The date when your load balancer was created.

$sel:location:LoadBalancer', loadBalancer_location - The AWS Region where your load balancer was created (e.g., us-east-2a). Lightsail automatically creates your load balancer across Availability Zones.

$sel:instancePort:LoadBalancer', loadBalancer_instancePort - The port where the load balancer will direct traffic to your Lightsail instances. For HTTP traffic, it's port 80. For HTTPS traffic, it's port 443.

$sel:configurationOptions:LoadBalancer', loadBalancer_configurationOptions - A string to string map of the configuration options for your load balancer. Valid values are listed below.

$sel:protocol:LoadBalancer', loadBalancer_protocol - The protocol you have enabled for your load balancer. Valid values are below.

You can't just have HTTP_HTTPS, but you can have just HTTP.

$sel:tlsCertificateSummaries:LoadBalancer', loadBalancer_tlsCertificateSummaries - An array of LoadBalancerTlsCertificateSummary objects that provide additional information about the SSL/TLS certificates. For example, if true, the certificate is attached to the load balancer.

$sel:name:LoadBalancer', loadBalancer_name - The name of the load balancer (e.g., my-load-balancer).

$sel:ipAddressType:LoadBalancer', loadBalancer_ipAddressType - The IP address type of the load balancer.

The possible values are ipv4 for IPv4 only, and dualstack for IPv4 and IPv6.

$sel:supportCode:LoadBalancer', loadBalancer_supportCode - The support code. Include this code in your email to support when you have questions about your Lightsail load balancer. This code enables our support team to look up your Lightsail information more easily.

$sel:publicPorts:LoadBalancer', loadBalancer_publicPorts - An array of public port settings for your load balancer. For HTTP, use port 80. For HTTPS, use port 443.

$sel:dnsName:LoadBalancer', loadBalancer_dnsName - The DNS name of your Lightsail load balancer.

$sel:instanceHealthSummary:LoadBalancer', loadBalancer_instanceHealthSummary - An array of InstanceHealthSummary objects describing the health of the load balancer.

$sel:tags:LoadBalancer', loadBalancer_tags - The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

loadBalancer_healthCheckPath :: Lens' LoadBalancer (Maybe Text)

The path you specified to perform your health checks. If no path is specified, the load balancer tries to make a request to the default (root) page.

loadBalancer_state :: Lens' LoadBalancer (Maybe LoadBalancerState)

The status of your load balancer. Valid values are below.

loadBalancer_resourceType :: Lens' LoadBalancer (Maybe ResourceType)

The resource type (e.g., LoadBalancer.

loadBalancer_arn :: Lens' LoadBalancer (Maybe Text)

The Amazon Resource Name (ARN) of the load balancer.

loadBalancer_createdAt :: Lens' LoadBalancer (Maybe UTCTime)

The date when your load balancer was created.

loadBalancer_location :: Lens' LoadBalancer (Maybe ResourceLocation)

The AWS Region where your load balancer was created (e.g., us-east-2a). Lightsail automatically creates your load balancer across Availability Zones.

loadBalancer_instancePort :: Lens' LoadBalancer (Maybe Int)

The port where the load balancer will direct traffic to your Lightsail instances. For HTTP traffic, it's port 80. For HTTPS traffic, it's port 443.

loadBalancer_configurationOptions :: Lens' LoadBalancer (Maybe (HashMap LoadBalancerAttributeName Text))

A string to string map of the configuration options for your load balancer. Valid values are listed below.

loadBalancer_protocol :: Lens' LoadBalancer (Maybe LoadBalancerProtocol)

The protocol you have enabled for your load balancer. Valid values are below.

You can't just have HTTP_HTTPS, but you can have just HTTP.

loadBalancer_tlsCertificateSummaries :: Lens' LoadBalancer (Maybe [LoadBalancerTlsCertificateSummary])

An array of LoadBalancerTlsCertificateSummary objects that provide additional information about the SSL/TLS certificates. For example, if true, the certificate is attached to the load balancer.

loadBalancer_name :: Lens' LoadBalancer (Maybe Text)

The name of the load balancer (e.g., my-load-balancer).

loadBalancer_ipAddressType :: Lens' LoadBalancer (Maybe IpAddressType)

The IP address type of the load balancer.

The possible values are ipv4 for IPv4 only, and dualstack for IPv4 and IPv6.

loadBalancer_supportCode :: Lens' LoadBalancer (Maybe Text)

The support code. Include this code in your email to support when you have questions about your Lightsail load balancer. This code enables our support team to look up your Lightsail information more easily.

loadBalancer_publicPorts :: Lens' LoadBalancer (Maybe [Int])

An array of public port settings for your load balancer. For HTTP, use port 80. For HTTPS, use port 443.

loadBalancer_dnsName :: Lens' LoadBalancer (Maybe Text)

The DNS name of your Lightsail load balancer.

loadBalancer_instanceHealthSummary :: Lens' LoadBalancer (Maybe [InstanceHealthSummary])

An array of InstanceHealthSummary objects describing the health of the load balancer.

loadBalancer_tags :: Lens' LoadBalancer (Maybe [Tag])

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.