Network.AWS.Lightsail.Types.LoadBalancerTlsCertificate

Description

Documentation

data LoadBalancerTlsCertificate

Describes a load balancer SSL/TLS certificate.

TLS is just an updated, more secure version of Secure Socket Layer (SSL).

See: newLoadBalancerTlsCertificate smart constructor.

Constructors

LoadBalancerTlsCertificate'

Fields

failureReason :: Maybe LoadBalancerTlsCertificateFailureReason

The validation failure reason, if any, of the certificate.

The following failure reasons are possible:

NO_AVAILABLE_CONTACTS - This failure applies to email validation, which is not available for Lightsail certificates.

ADDITIONAL_VERIFICATION_REQUIRED - Lightsail requires additional information to process this certificate request. This can happen as a fraud-protection measure, such as when the domain ranks within the Alexa top 1000 websites. To provide the required information, use the AWS Support Center to contact AWS Support.

You cannot request a certificate for Amazon-owned domain names such as those ending in amazonaws.com, cloudfront.net, or elasticbeanstalk.com.

DOMAIN_NOT_ALLOWED - One or more of the domain names in the certificate request was reported as an unsafe domain by VirusTotal. To correct the problem, search for your domain name on the VirusTotal website. If your domain is reported as suspicious, see Google Help for Hacked Websites to learn what you can do.

If you believe that the result is a false positive, notify the organization that is reporting the domain. VirusTotal is an aggregate of several antivirus and URL scanners and cannot remove your domain from a block list itself. After you correct the problem and the VirusTotal registry has been updated, request a new certificate.

If you see this error and your domain is not included in the VirusTotal list, visit the AWS Support Center and create a case.

INVALID_PUBLIC_DOMAIN - One or more of the domain names in the certificate request is not valid. Typically, this is because a domain name in the request is not a valid top-level domain. Try to request a certificate again, correcting any spelling errors or typos that were in the failed request, and ensure that all domain names in the request are for valid top-level domains. For example, you cannot request a certificate for example.invalidpublicdomain because invalidpublicdomain is not a valid top-level domain.

OTHER - Typically, this failure occurs when there is a typographical error in one or more of the domain names in the certificate request. Try to request a certificate again, correcting any spelling errors or typos that were in the failed request.

subject :: Maybe Text

The name of the entity that is associated with the public key contained in the certificate.

status :: Maybe LoadBalancerTlsCertificateStatus

The validation status of the SSL/TLS certificate. Valid values are below.

subjectAlternativeNames :: Maybe [Text]

An array of strings that specify the alternate domains (e.g., example2.com) and subdomains (e.g., blog.example.com) for the certificate.

resourceType :: Maybe ResourceType

The resource type (e.g., LoadBalancerTlsCertificate).

Instance - A Lightsail instance (a virtual private server)

StaticIp - A static IP address

KeyPair - The key pair used to connect to a Lightsail instance

InstanceSnapshot - A Lightsail instance snapshot

Domain - A DNS zone

PeeredVpc - A peered VPC

LoadBalancer - A Lightsail load balancer

LoadBalancerTlsCertificate - An SSL/TLS certificate associated with a Lightsail load balancer

Disk - A Lightsail block storage disk

DiskSnapshot - A block storage disk snapshot

arn :: Maybe Text

The Amazon Resource Name (ARN) of the SSL/TLS certificate.

createdAt :: Maybe POSIX

The time when you created your SSL/TLS certificate.

location :: Maybe ResourceLocation

The AWS Region and Availability Zone where you created your certificate.

loadBalancerName :: Maybe Text

The load balancer name where your SSL/TLS certificate is attached.

serial :: Maybe Text

The serial number of the certificate.

isAttached :: Maybe Bool

When true, the SSL/TLS certificate is attached to the Lightsail load balancer.

revokedAt :: Maybe POSIX

The timestamp when the certificate was revoked. This value is present only when the certificate status is REVOKED.

notBefore :: Maybe POSIX

The timestamp when the SSL/TLS certificate is first valid.

revocationReason :: Maybe LoadBalancerTlsCertificateRevocationReason

The reason the certificate was revoked. This value is present only when the certificate status is REVOKED.

domainName :: Maybe Text

The domain name for your SSL/TLS certificate.

name :: Maybe Text

The name of the SSL/TLS certificate (e.g., my-certificate).

renewalSummary :: Maybe LoadBalancerTlsCertificateRenewalSummary

An object that describes the status of the certificate renewal managed by Lightsail.

supportCode :: Maybe Text

The support code. Include this code in your email to support when you have questions about your Lightsail load balancer or SSL/TLS certificate. This code enables our support team to look up your Lightsail information more easily.

domainValidationRecords :: Maybe [LoadBalancerTlsCertificateDomainValidationRecord]

An array of LoadBalancerTlsCertificateDomainValidationRecord objects describing the records.

issuedAt :: Maybe POSIX

The time when the SSL/TLS certificate was issued.

keyAlgorithm :: Maybe Text

The algorithm used to generate the key pair (the public and private key).

signatureAlgorithm :: Maybe Text

The algorithm that was used to sign the certificate.

issuer :: Maybe Text

The issuer of the certificate.

tags :: Maybe [Tag]

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

notAfter :: Maybe POSIX

The timestamp when the SSL/TLS certificate expires.

newLoadBalancerTlsCertificate :: LoadBalancerTlsCertificate

Create a value of LoadBalancerTlsCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_failureReason - The validation failure reason, if any, of the certificate.

The following failure reasons are possible:

NO_AVAILABLE_CONTACTS - This failure applies to email validation, which is not available for Lightsail certificates.

ADDITIONAL_VERIFICATION_REQUIRED - Lightsail requires additional information to process this certificate request. This can happen as a fraud-protection measure, such as when the domain ranks within the Alexa top 1000 websites. To provide the required information, use the AWS Support Center to contact AWS Support.

You cannot request a certificate for Amazon-owned domain names such as those ending in amazonaws.com, cloudfront.net, or elasticbeanstalk.com.

DOMAIN_NOT_ALLOWED - One or more of the domain names in the certificate request was reported as an unsafe domain by VirusTotal. To correct the problem, search for your domain name on the VirusTotal website. If your domain is reported as suspicious, see Google Help for Hacked Websites to learn what you can do.

If you believe that the result is a false positive, notify the organization that is reporting the domain. VirusTotal is an aggregate of several antivirus and URL scanners and cannot remove your domain from a block list itself. After you correct the problem and the VirusTotal registry has been updated, request a new certificate.

If you see this error and your domain is not included in the VirusTotal list, visit the AWS Support Center and create a case.

INVALID_PUBLIC_DOMAIN - One or more of the domain names in the certificate request is not valid. Typically, this is because a domain name in the request is not a valid top-level domain. Try to request a certificate again, correcting any spelling errors or typos that were in the failed request, and ensure that all domain names in the request are for valid top-level domains. For example, you cannot request a certificate for example.invalidpublicdomain because invalidpublicdomain is not a valid top-level domain.

OTHER - Typically, this failure occurs when there is a typographical error in one or more of the domain names in the certificate request. Try to request a certificate again, correcting any spelling errors or typos that were in the failed request.

$sel:subject:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_subject - The name of the entity that is associated with the public key contained in the certificate.

$sel:status:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_status - The validation status of the SSL/TLS certificate. Valid values are below.

$sel:subjectAlternativeNames:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_subjectAlternativeNames - An array of strings that specify the alternate domains (e.g., example2.com) and subdomains (e.g., blog.example.com) for the certificate.

$sel:resourceType:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_resourceType - The resource type (e.g., LoadBalancerTlsCertificate).

Instance - A Lightsail instance (a virtual private server)

StaticIp - A static IP address

KeyPair - The key pair used to connect to a Lightsail instance

InstanceSnapshot - A Lightsail instance snapshot

Domain - A DNS zone

PeeredVpc - A peered VPC

LoadBalancer - A Lightsail load balancer

LoadBalancerTlsCertificate - An SSL/TLS certificate associated with a Lightsail load balancer

Disk - A Lightsail block storage disk

DiskSnapshot - A block storage disk snapshot

$sel:arn:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_arn - The Amazon Resource Name (ARN) of the SSL/TLS certificate.

$sel:createdAt:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_createdAt - The time when you created your SSL/TLS certificate.

$sel:location:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_location - The AWS Region and Availability Zone where you created your certificate.

$sel:loadBalancerName:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_loadBalancerName - The load balancer name where your SSL/TLS certificate is attached.

$sel:serial:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_serial - The serial number of the certificate.

$sel:isAttached:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_isAttached - When true, the SSL/TLS certificate is attached to the Lightsail load balancer.

$sel:revokedAt:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_revokedAt - The timestamp when the certificate was revoked. This value is present only when the certificate status is REVOKED.

$sel:notBefore:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_notBefore - The timestamp when the SSL/TLS certificate is first valid.

$sel:revocationReason:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_revocationReason - The reason the certificate was revoked. This value is present only when the certificate status is REVOKED.

$sel:domainName:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_domainName - The domain name for your SSL/TLS certificate.

$sel:name:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_name - The name of the SSL/TLS certificate (e.g., my-certificate).

$sel:renewalSummary:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_renewalSummary - An object that describes the status of the certificate renewal managed by Lightsail.

$sel:supportCode:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_supportCode - The support code. Include this code in your email to support when you have questions about your Lightsail load balancer or SSL/TLS certificate. This code enables our support team to look up your Lightsail information more easily.

$sel:domainValidationRecords:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_domainValidationRecords - An array of LoadBalancerTlsCertificateDomainValidationRecord objects describing the records.

$sel:issuedAt:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_issuedAt - The time when the SSL/TLS certificate was issued.

$sel:keyAlgorithm:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_keyAlgorithm - The algorithm used to generate the key pair (the public and private key).

$sel:signatureAlgorithm:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_signatureAlgorithm - The algorithm that was used to sign the certificate.

$sel:issuer:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_issuer - The issuer of the certificate.

$sel:tags:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_tags - The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

$sel:notAfter:LoadBalancerTlsCertificate', loadBalancerTlsCertificate_notAfter - The timestamp when the SSL/TLS certificate expires.

loadBalancerTlsCertificate_failureReason :: Lens' LoadBalancerTlsCertificate (Maybe LoadBalancerTlsCertificateFailureReason)

The validation failure reason, if any, of the certificate.

The following failure reasons are possible:

NO_AVAILABLE_CONTACTS - This failure applies to email validation, which is not available for Lightsail certificates.

ADDITIONAL_VERIFICATION_REQUIRED - Lightsail requires additional information to process this certificate request. This can happen as a fraud-protection measure, such as when the domain ranks within the Alexa top 1000 websites. To provide the required information, use the AWS Support Center to contact AWS Support.

You cannot request a certificate for Amazon-owned domain names such as those ending in amazonaws.com, cloudfront.net, or elasticbeanstalk.com.

DOMAIN_NOT_ALLOWED - One or more of the domain names in the certificate request was reported as an unsafe domain by VirusTotal. To correct the problem, search for your domain name on the VirusTotal website. If your domain is reported as suspicious, see Google Help for Hacked Websites to learn what you can do.

If you believe that the result is a false positive, notify the organization that is reporting the domain. VirusTotal is an aggregate of several antivirus and URL scanners and cannot remove your domain from a block list itself. After you correct the problem and the VirusTotal registry has been updated, request a new certificate.

If you see this error and your domain is not included in the VirusTotal list, visit the AWS Support Center and create a case.

INVALID_PUBLIC_DOMAIN - One or more of the domain names in the certificate request is not valid. Typically, this is because a domain name in the request is not a valid top-level domain. Try to request a certificate again, correcting any spelling errors or typos that were in the failed request, and ensure that all domain names in the request are for valid top-level domains. For example, you cannot request a certificate for example.invalidpublicdomain because invalidpublicdomain is not a valid top-level domain.

OTHER - Typically, this failure occurs when there is a typographical error in one or more of the domain names in the certificate request. Try to request a certificate again, correcting any spelling errors or typos that were in the failed request.

loadBalancerTlsCertificate_subject :: Lens' LoadBalancerTlsCertificate (Maybe Text)

The name of the entity that is associated with the public key contained in the certificate.

loadBalancerTlsCertificate_status :: Lens' LoadBalancerTlsCertificate (Maybe LoadBalancerTlsCertificateStatus)

The validation status of the SSL/TLS certificate. Valid values are below.

loadBalancerTlsCertificate_subjectAlternativeNames :: Lens' LoadBalancerTlsCertificate (Maybe [Text])

An array of strings that specify the alternate domains (e.g., example2.com) and subdomains (e.g., blog.example.com) for the certificate.

loadBalancerTlsCertificate_resourceType :: Lens' LoadBalancerTlsCertificate (Maybe ResourceType)

The resource type (e.g., LoadBalancerTlsCertificate).

Instance - A Lightsail instance (a virtual private server)

StaticIp - A static IP address

KeyPair - The key pair used to connect to a Lightsail instance

InstanceSnapshot - A Lightsail instance snapshot

Domain - A DNS zone

PeeredVpc - A peered VPC

LoadBalancer - A Lightsail load balancer

LoadBalancerTlsCertificate - An SSL/TLS certificate associated with a Lightsail load balancer

Disk - A Lightsail block storage disk

DiskSnapshot - A block storage disk snapshot

loadBalancerTlsCertificate_arn :: Lens' LoadBalancerTlsCertificate (Maybe Text)

The Amazon Resource Name (ARN) of the SSL/TLS certificate.

loadBalancerTlsCertificate_createdAt :: Lens' LoadBalancerTlsCertificate (Maybe UTCTime)

The time when you created your SSL/TLS certificate.

loadBalancerTlsCertificate_location :: Lens' LoadBalancerTlsCertificate (Maybe ResourceLocation)

The AWS Region and Availability Zone where you created your certificate.

loadBalancerTlsCertificate_loadBalancerName :: Lens' LoadBalancerTlsCertificate (Maybe Text)

The load balancer name where your SSL/TLS certificate is attached.

loadBalancerTlsCertificate_serial :: Lens' LoadBalancerTlsCertificate (Maybe Text)

The serial number of the certificate.

loadBalancerTlsCertificate_isAttached :: Lens' LoadBalancerTlsCertificate (Maybe Bool)

When true, the SSL/TLS certificate is attached to the Lightsail load balancer.

loadBalancerTlsCertificate_revokedAt :: Lens' LoadBalancerTlsCertificate (Maybe UTCTime)

The timestamp when the certificate was revoked. This value is present only when the certificate status is REVOKED.

loadBalancerTlsCertificate_notBefore :: Lens' LoadBalancerTlsCertificate (Maybe UTCTime)

The timestamp when the SSL/TLS certificate is first valid.

loadBalancerTlsCertificate_revocationReason :: Lens' LoadBalancerTlsCertificate (Maybe LoadBalancerTlsCertificateRevocationReason)

The reason the certificate was revoked. This value is present only when the certificate status is REVOKED.

loadBalancerTlsCertificate_domainName :: Lens' LoadBalancerTlsCertificate (Maybe Text)

The domain name for your SSL/TLS certificate.

loadBalancerTlsCertificate_name :: Lens' LoadBalancerTlsCertificate (Maybe Text)

The name of the SSL/TLS certificate (e.g., my-certificate).

loadBalancerTlsCertificate_renewalSummary :: Lens' LoadBalancerTlsCertificate (Maybe LoadBalancerTlsCertificateRenewalSummary)

An object that describes the status of the certificate renewal managed by Lightsail.

loadBalancerTlsCertificate_supportCode :: Lens' LoadBalancerTlsCertificate (Maybe Text)

The support code. Include this code in your email to support when you have questions about your Lightsail load balancer or SSL/TLS certificate. This code enables our support team to look up your Lightsail information more easily.

loadBalancerTlsCertificate_domainValidationRecords :: Lens' LoadBalancerTlsCertificate (Maybe [LoadBalancerTlsCertificateDomainValidationRecord])

An array of LoadBalancerTlsCertificateDomainValidationRecord objects describing the records.

loadBalancerTlsCertificate_issuedAt :: Lens' LoadBalancerTlsCertificate (Maybe UTCTime)

The time when the SSL/TLS certificate was issued.

loadBalancerTlsCertificate_keyAlgorithm :: Lens' LoadBalancerTlsCertificate (Maybe Text)

The algorithm used to generate the key pair (the public and private key).

loadBalancerTlsCertificate_signatureAlgorithm :: Lens' LoadBalancerTlsCertificate (Maybe Text)

The algorithm that was used to sign the certificate.

loadBalancerTlsCertificate_issuer :: Lens' LoadBalancerTlsCertificate (Maybe Text)

The issuer of the certificate.

loadBalancerTlsCertificate_tags :: Lens' LoadBalancerTlsCertificate (Maybe [Tag])

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

loadBalancerTlsCertificate_notAfter :: Lens' LoadBalancerTlsCertificate (Maybe UTCTime)

The timestamp when the SSL/TLS certificate expires.