Network.AWS.Lightsail.Types.LoadBalancerTlsCertificateRenewalSummary

Description

Documentation

data LoadBalancerTlsCertificateRenewalSummary

Contains information about the status of Lightsail's managed renewal for the certificate.

The renewal status of the certificate.

The following renewal status are possible:

PendingAutoRenewal - Lightsail is attempting to automatically validate the domain names in the certificate. No further action is required.

PendingValidation - Lightsail couldn't automatically validate one or more domain names in the certificate. You must take action to validate these domain names or the certificate won't be renewed. If you used DNS validation, check to make sure your certificate's domain validation records exist in your domain's DNS, and that your certificate remains in use.

Success - All domain names in the certificate are validated, and Lightsail renewed the certificate. No further action is required.

Failed - One or more domain names were not validated before the certificate expired, and Lightsail did not renew the certificate. You can request a new certificate using the CreateCertificate action.

See: newLoadBalancerTlsCertificateRenewalSummary smart constructor.

Constructors

LoadBalancerTlsCertificateRenewalSummary'

Fields

renewalStatus :: Maybe LoadBalancerTlsCertificateRenewalStatus

The renewal status of the certificate.

The following renewal status are possible:

PendingAutoRenewal - Lightsail is attempting to automatically validate the domain names of the certificate. No further action is required.

PendingValidation - Lightsail couldn't automatically validate one or more domain names of the certificate. You must take action to validate these domain names or the certificate won't be renewed. Check to make sure your certificate's domain validation records exist in your domain's DNS, and that your certificate remains in use.

Success - All domain names in the certificate are validated, and Lightsail renewed the certificate. No further action is required.

Failed - One or more domain names were not validated before the certificate expired, and Lightsail did not renew the certificate. You can request a new certificate using the CreateCertificate action.

domainValidationOptions :: Maybe [LoadBalancerTlsCertificateDomainValidationOption]

Contains information about the validation of each domain name in the certificate, as it pertains to Lightsail's managed renewal. This is different from the initial validation that occurs as a result of the RequestCertificate request.

newLoadBalancerTlsCertificateRenewalSummary :: LoadBalancerTlsCertificateRenewalSummary

Create a value of LoadBalancerTlsCertificateRenewalSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:renewalStatus:LoadBalancerTlsCertificateRenewalSummary', loadBalancerTlsCertificateRenewalSummary_renewalStatus - The renewal status of the certificate.

The following renewal status are possible:

PendingAutoRenewal - Lightsail is attempting to automatically validate the domain names of the certificate. No further action is required.

PendingValidation - Lightsail couldn't automatically validate one or more domain names of the certificate. You must take action to validate these domain names or the certificate won't be renewed. Check to make sure your certificate's domain validation records exist in your domain's DNS, and that your certificate remains in use.

Success - All domain names in the certificate are validated, and Lightsail renewed the certificate. No further action is required.

Failed - One or more domain names were not validated before the certificate expired, and Lightsail did not renew the certificate. You can request a new certificate using the CreateCertificate action.

$sel:domainValidationOptions:LoadBalancerTlsCertificateRenewalSummary', loadBalancerTlsCertificateRenewalSummary_domainValidationOptions - Contains information about the validation of each domain name in the certificate, as it pertains to Lightsail's managed renewal. This is different from the initial validation that occurs as a result of the RequestCertificate request.

loadBalancerTlsCertificateRenewalSummary_renewalStatus :: Lens' LoadBalancerTlsCertificateRenewalSummary (Maybe LoadBalancerTlsCertificateRenewalStatus)

The renewal status of the certificate.

The following renewal status are possible:

PendingAutoRenewal - Lightsail is attempting to automatically validate the domain names of the certificate. No further action is required.

PendingValidation - Lightsail couldn't automatically validate one or more domain names of the certificate. You must take action to validate these domain names or the certificate won't be renewed. Check to make sure your certificate's domain validation records exist in your domain's DNS, and that your certificate remains in use.

Success - All domain names in the certificate are validated, and Lightsail renewed the certificate. No further action is required.

Failed - One or more domain names were not validated before the certificate expired, and Lightsail did not renew the certificate. You can request a new certificate using the CreateCertificate action.

loadBalancerTlsCertificateRenewalSummary_domainValidationOptions :: Lens' LoadBalancerTlsCertificateRenewalSummary (Maybe [LoadBalancerTlsCertificateDomainValidationOption])

Contains information about the validation of each domain name in the certificate, as it pertains to Lightsail's managed renewal. This is different from the initial validation that occurs as a result of the RequestCertificate request.