Network.AWS.Lightsail.Types.LogEvent

Description

Documentation

data LogEvent

Describes a database log event.

See: newLogEvent smart constructor.

Constructors

LogEvent'

Fields

message :: Maybe Text

The message of the database log event.

createdAt :: Maybe POSIX

The timestamp when the database log event was created.

newLogEvent :: LogEvent

Create a value of LogEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:message:LogEvent', logEvent_message - The message of the database log event.

$sel:createdAt:LogEvent', logEvent_createdAt - The timestamp when the database log event was created.

logEvent_message :: Lens' LogEvent (Maybe Text)

The message of the database log event.

logEvent_createdAt :: Lens' LogEvent (Maybe UTCTime)

The timestamp when the database log event was created.