Network.AWS.Lightsail.Types.Operation

Description

Documentation

data Operation

Describes the API operation.

See: newOperation smart constructor.

Constructors

Operation'

Fields

operationDetails :: Maybe Text

Details about the operation (e.g., Debian-1GB-Ohio-1).

status :: Maybe OperationStatus

The status of the operation.

isTerminal :: Maybe Bool

A Boolean value indicating whether the operation is terminal.

id :: Maybe Text

The ID of the operation.

createdAt :: Maybe POSIX

The timestamp when the operation was initialized (e.g., 1479816991.349).

resourceType :: Maybe ResourceType

The resource type.

statusChangedAt :: Maybe POSIX

The timestamp when the status was changed (e.g., 1479816991.349).

errorCode :: Maybe Text

The error code.

operationType :: Maybe OperationType

The type of operation.

resourceName :: Maybe Text

The resource name.

location :: Maybe ResourceLocation

The AWS Region and Availability Zone.

errorDetails :: Maybe Text

The error details.

newOperation :: Operation

Create a value of Operation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operationDetails:Operation', operation_operationDetails - Details about the operation (e.g., Debian-1GB-Ohio-1).

$sel:status:Operation', operation_status - The status of the operation.

$sel:isTerminal:Operation', operation_isTerminal - A Boolean value indicating whether the operation is terminal.

$sel:id:Operation', operation_id - The ID of the operation.

$sel:createdAt:Operation', operation_createdAt - The timestamp when the operation was initialized (e.g., 1479816991.349).

$sel:resourceType:Operation', operation_resourceType - The resource type.

$sel:statusChangedAt:Operation', operation_statusChangedAt - The timestamp when the status was changed (e.g., 1479816991.349).

$sel:errorCode:Operation', operation_errorCode - The error code.

$sel:operationType:Operation', operation_operationType - The type of operation.

$sel:resourceName:Operation', operation_resourceName - The resource name.

$sel:location:Operation', operation_location - The AWS Region and Availability Zone.

$sel:errorDetails:Operation', operation_errorDetails - The error details.

operation_operationDetails :: Lens' Operation (Maybe Text)

Details about the operation (e.g., Debian-1GB-Ohio-1).

operation_status :: Lens' Operation (Maybe OperationStatus)

The status of the operation.

operation_isTerminal :: Lens' Operation (Maybe Bool)

A Boolean value indicating whether the operation is terminal.

operation_id :: Lens' Operation (Maybe Text)

The ID of the operation.

operation_createdAt :: Lens' Operation (Maybe UTCTime)

The timestamp when the operation was initialized (e.g., 1479816991.349).

operation_resourceType :: Lens' Operation (Maybe ResourceType)

The resource type.

operation_statusChangedAt :: Lens' Operation (Maybe UTCTime)

The timestamp when the status was changed (e.g., 1479816991.349).

operation_errorCode :: Lens' Operation (Maybe Text)

The error code.

operation_operationType :: Lens' Operation (Maybe OperationType)

The type of operation.

operation_resourceName :: Lens' Operation (Maybe Text)

The resource name.

operation_location :: Lens' Operation (Maybe ResourceLocation)

The AWS Region and Availability Zone.

operation_errorDetails :: Lens' Operation (Maybe Text)

The error details.