Network.AWS.Lightsail.Types.PendingMaintenanceAction

Description

Documentation

data PendingMaintenanceAction

Describes a pending database maintenance action.

See: newPendingMaintenanceAction smart constructor.

Constructors

PendingMaintenanceAction'

Fields

currentApplyDate :: Maybe POSIX

The effective date of the pending database maintenance action.

action :: Maybe Text

The type of pending database maintenance action.

description :: Maybe Text

Additional detail about the pending database maintenance action.

newPendingMaintenanceAction :: PendingMaintenanceAction

Create a value of PendingMaintenanceAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currentApplyDate:PendingMaintenanceAction', pendingMaintenanceAction_currentApplyDate - The effective date of the pending database maintenance action.

$sel:action:PendingMaintenanceAction', pendingMaintenanceAction_action - The type of pending database maintenance action.

$sel:description:PendingMaintenanceAction', pendingMaintenanceAction_description - Additional detail about the pending database maintenance action.

pendingMaintenanceAction_currentApplyDate :: Lens' PendingMaintenanceAction (Maybe UTCTime)

The effective date of the pending database maintenance action.

pendingMaintenanceAction_action :: Lens' PendingMaintenanceAction (Maybe Text)

The type of pending database maintenance action.

pendingMaintenanceAction_description :: Lens' PendingMaintenanceAction (Maybe Text)

Additional detail about the pending database maintenance action.