Network.AWS.Lightsail.Types.PendingModifiedRelationalDatabaseValues

Description

Documentation

data PendingModifiedRelationalDatabaseValues

Describes a pending database value modification.

See: newPendingModifiedRelationalDatabaseValues smart constructor.

Constructors

PendingModifiedRelationalDatabaseValues'

Fields

masterUserPassword :: Maybe Text

The password for the master user of the database.

backupRetentionEnabled :: Maybe Bool

A Boolean value indicating whether automated backup retention is enabled.

engineVersion :: Maybe Text

The database engine version.

newPendingModifiedRelationalDatabaseValues :: PendingModifiedRelationalDatabaseValues

Create a value of PendingModifiedRelationalDatabaseValues with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:masterUserPassword:PendingModifiedRelationalDatabaseValues', pendingModifiedRelationalDatabaseValues_masterUserPassword - The password for the master user of the database.

$sel:backupRetentionEnabled:PendingModifiedRelationalDatabaseValues', pendingModifiedRelationalDatabaseValues_backupRetentionEnabled - A Boolean value indicating whether automated backup retention is enabled.

$sel:engineVersion:PendingModifiedRelationalDatabaseValues', pendingModifiedRelationalDatabaseValues_engineVersion - The database engine version.

pendingModifiedRelationalDatabaseValues_masterUserPassword :: Lens' PendingModifiedRelationalDatabaseValues (Maybe Text)

The password for the master user of the database.

pendingModifiedRelationalDatabaseValues_backupRetentionEnabled :: Lens' PendingModifiedRelationalDatabaseValues (Maybe Bool)

A Boolean value indicating whether automated backup retention is enabled.

pendingModifiedRelationalDatabaseValues_engineVersion :: Lens' PendingModifiedRelationalDatabaseValues (Maybe Text)

The database engine version.