Network.AWS.Lightsail.Types.RelationalDatabaseBlueprint

Description

Documentation

data RelationalDatabaseBlueprint

Describes a database image, or blueprint. A blueprint describes the major engine version of a database.

See: newRelationalDatabaseBlueprint smart constructor.

Constructors

RelationalDatabaseBlueprint'

Fields

engineDescription :: Maybe Text

The description of the database engine for the database blueprint.

blueprintId :: Maybe Text

The ID for the database blueprint.

engineVersionDescription :: Maybe Text

The description of the database engine version for the database blueprint.

isEngineDefault :: Maybe Bool

A Boolean value indicating whether the engine version is the default for the database blueprint.

engineVersion :: Maybe Text

The database engine version for the database blueprint (for example, 5.7.23).

engine :: Maybe RelationalDatabaseEngine

The database software of the database blueprint (for example, MySQL).

newRelationalDatabaseBlueprint :: RelationalDatabaseBlueprint

Create a value of RelationalDatabaseBlueprint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineDescription:RelationalDatabaseBlueprint', relationalDatabaseBlueprint_engineDescription - The description of the database engine for the database blueprint.

$sel:blueprintId:RelationalDatabaseBlueprint', relationalDatabaseBlueprint_blueprintId - The ID for the database blueprint.

$sel:engineVersionDescription:RelationalDatabaseBlueprint', relationalDatabaseBlueprint_engineVersionDescription - The description of the database engine version for the database blueprint.

$sel:isEngineDefault:RelationalDatabaseBlueprint', relationalDatabaseBlueprint_isEngineDefault - A Boolean value indicating whether the engine version is the default for the database blueprint.

$sel:engineVersion:RelationalDatabaseBlueprint', relationalDatabaseBlueprint_engineVersion - The database engine version for the database blueprint (for example, 5.7.23).

$sel:engine:RelationalDatabaseBlueprint', relationalDatabaseBlueprint_engine - The database software of the database blueprint (for example, MySQL).

relationalDatabaseBlueprint_engineDescription :: Lens' RelationalDatabaseBlueprint (Maybe Text)

The description of the database engine for the database blueprint.

relationalDatabaseBlueprint_blueprintId :: Lens' RelationalDatabaseBlueprint (Maybe Text)

The ID for the database blueprint.

relationalDatabaseBlueprint_engineVersionDescription :: Lens' RelationalDatabaseBlueprint (Maybe Text)

The description of the database engine version for the database blueprint.

relationalDatabaseBlueprint_isEngineDefault :: Lens' RelationalDatabaseBlueprint (Maybe Bool)

A Boolean value indicating whether the engine version is the default for the database blueprint.

relationalDatabaseBlueprint_engineVersion :: Lens' RelationalDatabaseBlueprint (Maybe Text)

The database engine version for the database blueprint (for example, 5.7.23).

relationalDatabaseBlueprint_engine :: Lens' RelationalDatabaseBlueprint (Maybe RelationalDatabaseEngine)

The database software of the database blueprint (for example, MySQL).