Network.AWS.Lightsail.Types.RelationalDatabaseSnapshot

Description

Documentation

data RelationalDatabaseSnapshot

Describes a database snapshot.

See: newRelationalDatabaseSnapshot smart constructor.

Constructors

RelationalDatabaseSnapshot'

Fields

fromRelationalDatabaseName :: Maybe Text

The name of the source database from which the database snapshot was created.

createdAt :: Maybe POSIX

The timestamp when the database snapshot was created.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the database snapshot.

supportCode :: Maybe Text

The support code for the database snapshot. Include this code in your email to support when you have questions about a database snapshot in Lightsail. This code enables our support team to look up your Lightsail information more easily.

resourceType :: Maybe ResourceType

The Lightsail resource type.

sizeInGb :: Maybe Int

The size of the disk in GB (for example, 32) for the database snapshot.

state :: Maybe Text

The state of the database snapshot.

name :: Maybe Text

The name of the database snapshot.

engineVersion :: Maybe Text

The database engine version for the database snapshot (for example, 5.7.23).

tags :: Maybe [Tag]

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

fromRelationalDatabaseBundleId :: Maybe Text

The bundle ID of the database from which the database snapshot was created.

engine :: Maybe Text

The software of the database snapshot (for example, MySQL)

fromRelationalDatabaseArn :: Maybe Text

The Amazon Resource Name (ARN) of the database from which the database snapshot was created.

location :: Maybe ResourceLocation

The Region name and Availability Zone where the database snapshot is located.

fromRelationalDatabaseBlueprintId :: Maybe Text

The blueprint ID of the database from which the database snapshot was created. A blueprint describes the major engine version of a database.

newRelationalDatabaseSnapshot :: RelationalDatabaseSnapshot

Create a value of RelationalDatabaseSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fromRelationalDatabaseName:RelationalDatabaseSnapshot', relationalDatabaseSnapshot_fromRelationalDatabaseName - The name of the source database from which the database snapshot was created.

$sel:createdAt:RelationalDatabaseSnapshot', relationalDatabaseSnapshot_createdAt - The timestamp when the database snapshot was created.

$sel:arn:RelationalDatabaseSnapshot', relationalDatabaseSnapshot_arn - The Amazon Resource Name (ARN) of the database snapshot.

$sel:supportCode:RelationalDatabaseSnapshot', relationalDatabaseSnapshot_supportCode - The support code for the database snapshot. Include this code in your email to support when you have questions about a database snapshot in Lightsail. This code enables our support team to look up your Lightsail information more easily.

$sel:resourceType:RelationalDatabaseSnapshot', relationalDatabaseSnapshot_resourceType - The Lightsail resource type.

$sel:sizeInGb:RelationalDatabaseSnapshot', relationalDatabaseSnapshot_sizeInGb - The size of the disk in GB (for example, 32) for the database snapshot.

$sel:state:RelationalDatabaseSnapshot', relationalDatabaseSnapshot_state - The state of the database snapshot.

$sel:name:RelationalDatabaseSnapshot', relationalDatabaseSnapshot_name - The name of the database snapshot.

$sel:engineVersion:RelationalDatabaseSnapshot', relationalDatabaseSnapshot_engineVersion - The database engine version for the database snapshot (for example, 5.7.23).

$sel:tags:RelationalDatabaseSnapshot', relationalDatabaseSnapshot_tags - The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

$sel:fromRelationalDatabaseBundleId:RelationalDatabaseSnapshot', relationalDatabaseSnapshot_fromRelationalDatabaseBundleId - The bundle ID of the database from which the database snapshot was created.

$sel:engine:RelationalDatabaseSnapshot', relationalDatabaseSnapshot_engine - The software of the database snapshot (for example, MySQL)

$sel:fromRelationalDatabaseArn:RelationalDatabaseSnapshot', relationalDatabaseSnapshot_fromRelationalDatabaseArn - The Amazon Resource Name (ARN) of the database from which the database snapshot was created.

$sel:location:RelationalDatabaseSnapshot', relationalDatabaseSnapshot_location - The Region name and Availability Zone where the database snapshot is located.

$sel:fromRelationalDatabaseBlueprintId:RelationalDatabaseSnapshot', relationalDatabaseSnapshot_fromRelationalDatabaseBlueprintId - The blueprint ID of the database from which the database snapshot was created. A blueprint describes the major engine version of a database.

relationalDatabaseSnapshot_fromRelationalDatabaseName :: Lens' RelationalDatabaseSnapshot (Maybe Text)

The name of the source database from which the database snapshot was created.

relationalDatabaseSnapshot_createdAt :: Lens' RelationalDatabaseSnapshot (Maybe UTCTime)

The timestamp when the database snapshot was created.

relationalDatabaseSnapshot_arn :: Lens' RelationalDatabaseSnapshot (Maybe Text)

The Amazon Resource Name (ARN) of the database snapshot.

relationalDatabaseSnapshot_supportCode :: Lens' RelationalDatabaseSnapshot (Maybe Text)

The support code for the database snapshot. Include this code in your email to support when you have questions about a database snapshot in Lightsail. This code enables our support team to look up your Lightsail information more easily.

relationalDatabaseSnapshot_resourceType :: Lens' RelationalDatabaseSnapshot (Maybe ResourceType)

The Lightsail resource type.

relationalDatabaseSnapshot_sizeInGb :: Lens' RelationalDatabaseSnapshot (Maybe Int)

The size of the disk in GB (for example, 32) for the database snapshot.

relationalDatabaseSnapshot_state :: Lens' RelationalDatabaseSnapshot (Maybe Text)

The state of the database snapshot.

relationalDatabaseSnapshot_name :: Lens' RelationalDatabaseSnapshot (Maybe Text)

The name of the database snapshot.

relationalDatabaseSnapshot_engineVersion :: Lens' RelationalDatabaseSnapshot (Maybe Text)

The database engine version for the database snapshot (for example, 5.7.23).

relationalDatabaseSnapshot_tags :: Lens' RelationalDatabaseSnapshot (Maybe [Tag])

The tag keys and optional values for the resource. For more information about tags in Lightsail, see the Amazon Lightsail Developer Guide.

relationalDatabaseSnapshot_fromRelationalDatabaseBundleId :: Lens' RelationalDatabaseSnapshot (Maybe Text)

The bundle ID of the database from which the database snapshot was created.

relationalDatabaseSnapshot_engine :: Lens' RelationalDatabaseSnapshot (Maybe Text)

The software of the database snapshot (for example, MySQL)

relationalDatabaseSnapshot_fromRelationalDatabaseArn :: Lens' RelationalDatabaseSnapshot (Maybe Text)

The Amazon Resource Name (ARN) of the database from which the database snapshot was created.

relationalDatabaseSnapshot_location :: Lens' RelationalDatabaseSnapshot (Maybe ResourceLocation)

The Region name and Availability Zone where the database snapshot is located.

relationalDatabaseSnapshot_fromRelationalDatabaseBlueprintId :: Lens' RelationalDatabaseSnapshot (Maybe Text)

The blueprint ID of the database from which the database snapshot was created. A blueprint describes the major engine version of a database.