Network.AWS.Lightsail.Types.StaticIp

Description

Documentation

data StaticIp

Describes a static IP.

See: newStaticIp smart constructor.

Constructors

StaticIp'

Fields

isAttached :: Maybe Bool

A Boolean value indicating whether the static IP is attached.

createdAt :: Maybe POSIX

The timestamp when the static IP was created (e.g., 1479735304.222).

arn :: Maybe Text

The Amazon Resource Name (ARN) of the static IP (e.g., arn:aws:lightsail:us-east-2:123456789101:StaticIp/9cbb4a9e-f8e3-4dfe-b57e-12345EXAMPLE).

supportCode :: Maybe Text

The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

resourceType :: Maybe ResourceType

The resource type (usually StaticIp).

ipAddress :: Maybe Text

The static IP address.

name :: Maybe Text

The name of the static IP (e.g., StaticIP-Ohio-EXAMPLE).

attachedTo :: Maybe Text

The instance where the static IP is attached (e.g., Amazon_Linux-1GB-Ohio-1).

location :: Maybe ResourceLocation

The region and Availability Zone where the static IP was created.

newStaticIp :: StaticIp

Create a value of StaticIp with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isAttached:StaticIp', staticIp_isAttached - A Boolean value indicating whether the static IP is attached.

$sel:createdAt:StaticIp', staticIp_createdAt - The timestamp when the static IP was created (e.g., 1479735304.222).

$sel:arn:StaticIp', staticIp_arn - The Amazon Resource Name (ARN) of the static IP (e.g., arn:aws:lightsail:us-east-2:123456789101:StaticIp/9cbb4a9e-f8e3-4dfe-b57e-12345EXAMPLE).

$sel:supportCode:StaticIp', staticIp_supportCode - The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

$sel:resourceType:StaticIp', staticIp_resourceType - The resource type (usually StaticIp).

$sel:ipAddress:StaticIp', staticIp_ipAddress - The static IP address.

$sel:name:StaticIp', staticIp_name - The name of the static IP (e.g., StaticIP-Ohio-EXAMPLE).

$sel:attachedTo:StaticIp', staticIp_attachedTo - The instance where the static IP is attached (e.g., Amazon_Linux-1GB-Ohio-1).

$sel:location:StaticIp', staticIp_location - The region and Availability Zone where the static IP was created.

staticIp_isAttached :: Lens' StaticIp (Maybe Bool)

A Boolean value indicating whether the static IP is attached.

staticIp_createdAt :: Lens' StaticIp (Maybe UTCTime)

The timestamp when the static IP was created (e.g., 1479735304.222).

staticIp_arn :: Lens' StaticIp (Maybe Text)

The Amazon Resource Name (ARN) of the static IP (e.g., arn:aws:lightsail:us-east-2:123456789101:StaticIp/9cbb4a9e-f8e3-4dfe-b57e-12345EXAMPLE).

staticIp_supportCode :: Lens' StaticIp (Maybe Text)

The support code. Include this code in your email to support when you have questions about an instance or another resource in Lightsail. This code enables our support team to look up your Lightsail information more easily.

staticIp_resourceType :: Lens' StaticIp (Maybe ResourceType)

The resource type (usually StaticIp).

staticIp_ipAddress :: Lens' StaticIp (Maybe Text)

The static IP address.

staticIp_name :: Lens' StaticIp (Maybe Text)

The name of the static IP (e.g., StaticIP-Ohio-EXAMPLE).

staticIp_attachedTo :: Lens' StaticIp (Maybe Text)

The instance where the static IP is attached (e.g., Amazon_Linux-1GB-Ohio-1).

staticIp_location :: Lens' StaticIp (Maybe ResourceLocation)

The region and Availability Zone where the static IP was created.