Network.AWS.Lightsail.UpdateBucket

Description

Updates an existing Amazon Lightsail bucket.

Use this action to update the configuration of an existing bucket, such as versioning, public accessibility, and the AWS accounts that can access the bucket.

Creating a Request

data UpdateBucket

See: newUpdateBucket smart constructor.

Constructors

UpdateBucket'

Fields

readonlyAccessAccounts :: Maybe [Text]

An array of strings to specify the AWS account IDs that can access the bucket.

You can give a maximum of 10 AWS accounts access to a bucket.

accessRules :: Maybe AccessRules

An object that sets the public accessibility of objects in the specified bucket.

versioning :: Maybe Text

Specifies whether to enable or suspend versioning of objects in the bucket.

The following options can be specified:

Enabled - Enables versioning of objects in the specified bucket.

Suspended - Suspends versioning of objects in the specified bucket. Existing object versions are retained.

bucketName :: Text

The name of the bucket to update.

newUpdateBucket

Arguments

:: Text
$sel:bucketName:UpdateBucket'

-> UpdateBucket

Create a value of UpdateBucket with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:readonlyAccessAccounts:UpdateBucket', updateBucket_readonlyAccessAccounts - An array of strings to specify the AWS account IDs that can access the bucket.

You can give a maximum of 10 AWS accounts access to a bucket.

$sel:accessRules:UpdateBucket', updateBucket_accessRules - An object that sets the public accessibility of objects in the specified bucket.

$sel:versioning:UpdateBucket', updateBucket_versioning - Specifies whether to enable or suspend versioning of objects in the bucket.

The following options can be specified:

Enabled - Enables versioning of objects in the specified bucket.

Suspended - Suspends versioning of objects in the specified bucket. Existing object versions are retained.

$sel:bucketName:UpdateBucket', updateBucket_bucketName - The name of the bucket to update.

Request Lenses

updateBucket_readonlyAccessAccounts :: Lens' UpdateBucket (Maybe [Text])

An array of strings to specify the AWS account IDs that can access the bucket.

You can give a maximum of 10 AWS accounts access to a bucket.

updateBucket_accessRules :: Lens' UpdateBucket (Maybe AccessRules)

An object that sets the public accessibility of objects in the specified bucket.

updateBucket_versioning :: Lens' UpdateBucket (Maybe Text)

Specifies whether to enable or suspend versioning of objects in the bucket.

The following options can be specified:

Enabled - Enables versioning of objects in the specified bucket.

Suspended - Suspends versioning of objects in the specified bucket. Existing object versions are retained.

updateBucket_bucketName :: Lens' UpdateBucket Text

The name of the bucket to update.

Destructuring the Response

data UpdateBucketResponse

See: newUpdateBucketResponse smart constructor.

Constructors

UpdateBucketResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

bucket :: Maybe Bucket

An object that describes the bucket that is updated.

httpStatus :: Int

The response's http status code.

newUpdateBucketResponse

Arguments

:: Int
$sel:httpStatus:UpdateBucketResponse'

-> UpdateBucketResponse

Create a value of UpdateBucketResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:UpdateBucketResponse', updateBucketResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:bucket:UpdateBucketResponse', updateBucketResponse_bucket - An object that describes the bucket that is updated.

$sel:httpStatus:UpdateBucketResponse', updateBucketResponse_httpStatus - The response's http status code.

Response Lenses

updateBucketResponse_operations :: Lens' UpdateBucketResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

updateBucketResponse_bucket :: Lens' UpdateBucketResponse (Maybe Bucket)

An object that describes the bucket that is updated.

updateBucketResponse_httpStatus :: Lens' UpdateBucketResponse Int

The response's http status code.