Network.AWS.Lightsail.UpdateContainerService

Description

Updates the configuration of your Amazon Lightsail container service, such as its power, scale, and public domain names.

Creating a Request

data UpdateContainerService

See: newUpdateContainerService smart constructor.

Constructors

UpdateContainerService'

Fields

power :: Maybe ContainerServicePowerName

The power for the container service.

The power specifies the amount of memory, vCPUs, and base monthly cost of each node of the container service. The power and scale of a container service makes up its configured capacity. To determine the monthly price of your container service, multiply the base price of the power with the scale (the number of nodes) of the service.

Use the GetContainerServicePowers action to view the specifications of each power option.

scale :: Maybe Natural

The scale for the container service.

The scale specifies the allocated compute nodes of the container service. The power and scale of a container service makes up its configured capacity. To determine the monthly price of your container service, multiply the base price of the power with the scale (the number of nodes) of the service.

publicDomainNames :: Maybe (HashMap Text [Text])

The public domain names to use with the container service, such as example.com and www.example.com.

You can specify up to four public domain names for a container service. The domain names that you specify are used when you create a deployment with a container configured as the public endpoint of your container service.

If you don't specify public domain names, then you can use the default domain of the container service.

You must create and validate an SSL/TLS certificate before you can use public domain names with your container service. Use the CreateCertificate action to create a certificate for the public domain names you want to use with your container service.

You can specify public domain names using a string to array map as shown in the example later on this page.

isDisabled :: Maybe Bool

A Boolean value to indicate whether the container service is disabled.

serviceName :: Text

The name of the container service to update.

newUpdateContainerService

Arguments

:: Text
$sel:serviceName:UpdateContainerService'

-> UpdateContainerService

Create a value of UpdateContainerService with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:power:UpdateContainerService', updateContainerService_power - The power for the container service.

The power specifies the amount of memory, vCPUs, and base monthly cost of each node of the container service. The power and scale of a container service makes up its configured capacity. To determine the monthly price of your container service, multiply the base price of the power with the scale (the number of nodes) of the service.

Use the GetContainerServicePowers action to view the specifications of each power option.

$sel:scale:UpdateContainerService', updateContainerService_scale - The scale for the container service.

The scale specifies the allocated compute nodes of the container service. The power and scale of a container service makes up its configured capacity. To determine the monthly price of your container service, multiply the base price of the power with the scale (the number of nodes) of the service.

$sel:publicDomainNames:UpdateContainerService', updateContainerService_publicDomainNames - The public domain names to use with the container service, such as example.com and www.example.com.

You can specify up to four public domain names for a container service. The domain names that you specify are used when you create a deployment with a container configured as the public endpoint of your container service.

If you don't specify public domain names, then you can use the default domain of the container service.

You must create and validate an SSL/TLS certificate before you can use public domain names with your container service. Use the CreateCertificate action to create a certificate for the public domain names you want to use with your container service.

You can specify public domain names using a string to array map as shown in the example later on this page.

$sel:isDisabled:UpdateContainerService', updateContainerService_isDisabled - A Boolean value to indicate whether the container service is disabled.

$sel:serviceName:UpdateContainerService', updateContainerService_serviceName - The name of the container service to update.

Request Lenses

updateContainerService_power :: Lens' UpdateContainerService (Maybe ContainerServicePowerName)

The power for the container service.

The power specifies the amount of memory, vCPUs, and base monthly cost of each node of the container service. The power and scale of a container service makes up its configured capacity. To determine the monthly price of your container service, multiply the base price of the power with the scale (the number of nodes) of the service.

Use the GetContainerServicePowers action to view the specifications of each power option.

updateContainerService_scale :: Lens' UpdateContainerService (Maybe Natural)

The scale for the container service.

The scale specifies the allocated compute nodes of the container service. The power and scale of a container service makes up its configured capacity. To determine the monthly price of your container service, multiply the base price of the power with the scale (the number of nodes) of the service.

updateContainerService_publicDomainNames :: Lens' UpdateContainerService (Maybe (HashMap Text [Text]))

The public domain names to use with the container service, such as example.com and www.example.com.

You can specify up to four public domain names for a container service. The domain names that you specify are used when you create a deployment with a container configured as the public endpoint of your container service.

If you don't specify public domain names, then you can use the default domain of the container service.

You must create and validate an SSL/TLS certificate before you can use public domain names with your container service. Use the CreateCertificate action to create a certificate for the public domain names you want to use with your container service.

You can specify public domain names using a string to array map as shown in the example later on this page.

updateContainerService_isDisabled :: Lens' UpdateContainerService (Maybe Bool)

A Boolean value to indicate whether the container service is disabled.

updateContainerService_serviceName :: Lens' UpdateContainerService Text

The name of the container service to update.

Destructuring the Response

data UpdateContainerServiceResponse

See: newUpdateContainerServiceResponse smart constructor.

Constructors

UpdateContainerServiceResponse'

Fields

containerService :: Maybe ContainerService

An object that describes a container service.

httpStatus :: Int

The response's http status code.

newUpdateContainerServiceResponse

Arguments

:: Int
$sel:httpStatus:UpdateContainerServiceResponse'

-> UpdateContainerServiceResponse

Create a value of UpdateContainerServiceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerService:UpdateContainerServiceResponse', updateContainerServiceResponse_containerService - An object that describes a container service.

$sel:httpStatus:UpdateContainerServiceResponse', updateContainerServiceResponse_httpStatus - The response's http status code.

Response Lenses

updateContainerServiceResponse_containerService :: Lens' UpdateContainerServiceResponse (Maybe ContainerService)

An object that describes a container service.

updateContainerServiceResponse_httpStatus :: Lens' UpdateContainerServiceResponse Int

The response's http status code.