Network.AWS.Lightsail.UpdateDistribution

Description

Updates an existing Amazon Lightsail content delivery network (CDN) distribution.

Use this action to update the configuration of your existing distribution.

Creating a Request

data UpdateDistribution

See: newUpdateDistribution smart constructor.

Constructors

UpdateDistribution'

Fields

isEnabled :: Maybe Bool

Indicates whether to enable the distribution.

origin :: Maybe InputOrigin

An object that describes the origin resource for the distribution, such as a Lightsail instance or load balancer.

The distribution pulls, caches, and serves content from the origin.

cacheBehaviorSettings :: Maybe CacheSettings

An object that describes the cache behavior settings for the distribution.

The cacheBehaviorSettings specified in your UpdateDistributionRequest will replace your distribution's existing settings.

cacheBehaviors :: Maybe [CacheBehaviorPerPath]

An array of objects that describe the per-path cache behavior for the distribution.

defaultCacheBehavior :: Maybe CacheBehavior

An object that describes the default cache behavior for the distribution.

distributionName :: Text

The name of the distribution to update.

Use the GetDistributions action to get a list of distribution names that you can specify.

newUpdateDistribution

Arguments

:: Text
$sel:distributionName:UpdateDistribution'

-> UpdateDistribution

Create a value of UpdateDistribution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isEnabled:UpdateDistribution', updateDistribution_isEnabled - Indicates whether to enable the distribution.

$sel:origin:UpdateDistribution', updateDistribution_origin - An object that describes the origin resource for the distribution, such as a Lightsail instance or load balancer.

The distribution pulls, caches, and serves content from the origin.

$sel:cacheBehaviorSettings:UpdateDistribution', updateDistribution_cacheBehaviorSettings - An object that describes the cache behavior settings for the distribution.

The cacheBehaviorSettings specified in your UpdateDistributionRequest will replace your distribution's existing settings.

$sel:cacheBehaviors:UpdateDistribution', updateDistribution_cacheBehaviors - An array of objects that describe the per-path cache behavior for the distribution.

$sel:defaultCacheBehavior:UpdateDistribution', updateDistribution_defaultCacheBehavior - An object that describes the default cache behavior for the distribution.

$sel:distributionName:UpdateDistribution', updateDistribution_distributionName - The name of the distribution to update.

Use the GetDistributions action to get a list of distribution names that you can specify.

Request Lenses

updateDistribution_isEnabled :: Lens' UpdateDistribution (Maybe Bool)

Indicates whether to enable the distribution.

updateDistribution_origin :: Lens' UpdateDistribution (Maybe InputOrigin)

An object that describes the origin resource for the distribution, such as a Lightsail instance or load balancer.

The distribution pulls, caches, and serves content from the origin.

updateDistribution_cacheBehaviorSettings :: Lens' UpdateDistribution (Maybe CacheSettings)

An object that describes the cache behavior settings for the distribution.

The cacheBehaviorSettings specified in your UpdateDistributionRequest will replace your distribution's existing settings.

updateDistribution_cacheBehaviors :: Lens' UpdateDistribution (Maybe [CacheBehaviorPerPath])

An array of objects that describe the per-path cache behavior for the distribution.

updateDistribution_defaultCacheBehavior :: Lens' UpdateDistribution (Maybe CacheBehavior)

An object that describes the default cache behavior for the distribution.

updateDistribution_distributionName :: Lens' UpdateDistribution Text

The name of the distribution to update.

Use the GetDistributions action to get a list of distribution names that you can specify.

Destructuring the Response

data UpdateDistributionResponse

See: newUpdateDistributionResponse smart constructor.

Constructors

UpdateDistributionResponse'

Fields

operation :: Maybe Operation

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newUpdateDistributionResponse

Arguments

:: Int
$sel:httpStatus:UpdateDistributionResponse'

-> UpdateDistributionResponse

Create a value of UpdateDistributionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operation:UpdateDistributionResponse', updateDistributionResponse_operation - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:UpdateDistributionResponse', updateDistributionResponse_httpStatus - The response's http status code.

Response Lenses

updateDistributionResponse_operation :: Lens' UpdateDistributionResponse (Maybe Operation)

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

updateDistributionResponse_httpStatus :: Lens' UpdateDistributionResponse Int

The response's http status code.