Network.AWS.Lightsail.UpdateLoadBalancerAttribute

Description

Updates the specified attribute for a load balancer. You can only update one attribute at a time.

The update load balancer attribute operation supports tag-based access control via resource tags applied to the resource identified by load balancer name. For more information, see the Amazon Lightsail Developer Guide.

Creating a Request

data UpdateLoadBalancerAttribute

See: newUpdateLoadBalancerAttribute smart constructor.

Constructors

UpdateLoadBalancerAttribute'

Fields

loadBalancerName :: Text

The name of the load balancer that you want to modify (e.g., my-load-balancer.

attributeName :: LoadBalancerAttributeName

The name of the attribute you want to update. Valid values are below.

attributeValue :: Text

The value that you want to specify for the attribute name.

newUpdateLoadBalancerAttribute

Arguments

:: Text
$sel:loadBalancerName:UpdateLoadBalancerAttribute'

-> LoadBalancerAttributeName
$sel:attributeName:UpdateLoadBalancerAttribute'

-> Text
$sel:attributeValue:UpdateLoadBalancerAttribute'

-> UpdateLoadBalancerAttribute

Create a value of UpdateLoadBalancerAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loadBalancerName:UpdateLoadBalancerAttribute', updateLoadBalancerAttribute_loadBalancerName - The name of the load balancer that you want to modify (e.g., my-load-balancer.

$sel:attributeName:UpdateLoadBalancerAttribute', updateLoadBalancerAttribute_attributeName - The name of the attribute you want to update. Valid values are below.

$sel:attributeValue:UpdateLoadBalancerAttribute', updateLoadBalancerAttribute_attributeValue - The value that you want to specify for the attribute name.

Request Lenses

updateLoadBalancerAttribute_loadBalancerName :: Lens' UpdateLoadBalancerAttribute Text

The name of the load balancer that you want to modify (e.g., my-load-balancer.

updateLoadBalancerAttribute_attributeName :: Lens' UpdateLoadBalancerAttribute LoadBalancerAttributeName

The name of the attribute you want to update. Valid values are below.

updateLoadBalancerAttribute_attributeValue :: Lens' UpdateLoadBalancerAttribute Text

The value that you want to specify for the attribute name.

Destructuring the Response

data UpdateLoadBalancerAttributeResponse

See: newUpdateLoadBalancerAttributeResponse smart constructor.

Constructors

UpdateLoadBalancerAttributeResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newUpdateLoadBalancerAttributeResponse

Arguments

:: Int
$sel:httpStatus:UpdateLoadBalancerAttributeResponse'

-> UpdateLoadBalancerAttributeResponse

Create a value of UpdateLoadBalancerAttributeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:UpdateLoadBalancerAttributeResponse', updateLoadBalancerAttributeResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:UpdateLoadBalancerAttributeResponse', updateLoadBalancerAttributeResponse_httpStatus - The response's http status code.

Response Lenses

updateLoadBalancerAttributeResponse_operations :: Lens' UpdateLoadBalancerAttributeResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

updateLoadBalancerAttributeResponse_httpStatus :: Lens' UpdateLoadBalancerAttributeResponse Int

The response's http status code.