Network.AWS.Lightsail.UpdateRelationalDatabase

Description

Allows the update of one or more attributes of a database in Amazon Lightsail.

Updates are applied immediately, or in cases where the updates could result in an outage, are applied during the database's predefined maintenance window.

The update relational database operation supports tag-based access control via resource tags applied to the resource identified by relationalDatabaseName. For more information, see the Amazon Lightsail Developer Guide.

Creating a Request

data UpdateRelationalDatabase

See: newUpdateRelationalDatabase smart constructor.

Constructors

UpdateRelationalDatabase'

Fields

masterUserPassword :: Maybe (Sensitive Text)

The password for the master user. The password can include any printable ASCII character except "/", """, or "@".

MySQL

Constraints: Must contain from 8 to 41 characters.

PostgreSQL

Constraints: Must contain from 8 to 128 characters.

publiclyAccessible :: Maybe Bool

Specifies the accessibility options for your database. A value of true specifies a database that is available to resources outside of your Lightsail account. A value of false specifies a database that is available only to your Lightsail resources in the same region as your database.

enableBackupRetention :: Maybe Bool

When true, enables automated backup retention for your database.

Updates are applied during the next maintenance window because this can result in an outage.

preferredMaintenanceWindow :: Maybe Text

The weekly time range during which system maintenance can occur on your database.

The default is a 30-minute window selected at random from an 8-hour block of time for each AWS Region, occurring on a random day of the week.

Constraints:

Must be in the ddd:hh24:mi-ddd:hh24:mi format.

Valid days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Must be at least 30 minutes.

Specified in Coordinated Universal Time (UTC).

Example: Tue:17:00-Tue:17:30

caCertificateIdentifier :: Maybe Text

Indicates the certificate that needs to be associated with the database.

preferredBackupWindow :: Maybe Text

The daily time range during which automated backups are created for your database if automated backups are enabled.

Constraints:

Must be in the hh24:mi-hh24:mi format.

Example: 16:00-16:30

Specified in Coordinated Universal Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

applyImmediately :: Maybe Bool

When true, applies changes immediately. When false, applies changes during the preferred maintenance window. Some changes may cause an outage.

Default: false

rotateMasterUserPassword :: Maybe Bool

When true, the master user password is changed to a new strong password generated by Lightsail.

Use the get relational database master user password operation to get the new password.

disableBackupRetention :: Maybe Bool

When true, disables automated backup retention for your database.

Disabling backup retention deletes all automated database backups. Before disabling this, you may want to create a snapshot of your database using the create relational database snapshot operation.

Updates are applied during the next maintenance window because this can result in an outage.

relationalDatabaseName :: Text

The name of your Lightsail database resource to update.

newUpdateRelationalDatabase

Arguments

:: Text
$sel:relationalDatabaseName:UpdateRelationalDatabase'

-> UpdateRelationalDatabase

Create a value of UpdateRelationalDatabase with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:masterUserPassword:UpdateRelationalDatabase', updateRelationalDatabase_masterUserPassword - The password for the master user. The password can include any printable ASCII character except "/", """, or "@".

MySQL

Constraints: Must contain from 8 to 41 characters.

PostgreSQL

Constraints: Must contain from 8 to 128 characters.

$sel:publiclyAccessible:UpdateRelationalDatabase', updateRelationalDatabase_publiclyAccessible - Specifies the accessibility options for your database. A value of true specifies a database that is available to resources outside of your Lightsail account. A value of false specifies a database that is available only to your Lightsail resources in the same region as your database.

$sel:enableBackupRetention:UpdateRelationalDatabase', updateRelationalDatabase_enableBackupRetention - When true, enables automated backup retention for your database.

Updates are applied during the next maintenance window because this can result in an outage.

$sel:preferredMaintenanceWindow:UpdateRelationalDatabase', updateRelationalDatabase_preferredMaintenanceWindow - The weekly time range during which system maintenance can occur on your database.

The default is a 30-minute window selected at random from an 8-hour block of time for each AWS Region, occurring on a random day of the week.

Constraints:

Must be in the ddd:hh24:mi-ddd:hh24:mi format.

Valid days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Must be at least 30 minutes.

Specified in Coordinated Universal Time (UTC).

Example: Tue:17:00-Tue:17:30

$sel:caCertificateIdentifier:UpdateRelationalDatabase', updateRelationalDatabase_caCertificateIdentifier - Indicates the certificate that needs to be associated with the database.

$sel:preferredBackupWindow:UpdateRelationalDatabase', updateRelationalDatabase_preferredBackupWindow - The daily time range during which automated backups are created for your database if automated backups are enabled.

Constraints:

Must be in the hh24:mi-hh24:mi format.

Example: 16:00-16:30

Specified in Coordinated Universal Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

$sel:applyImmediately:UpdateRelationalDatabase', updateRelationalDatabase_applyImmediately - When true, applies changes immediately. When false, applies changes during the preferred maintenance window. Some changes may cause an outage.

Default: false

$sel:rotateMasterUserPassword:UpdateRelationalDatabase', updateRelationalDatabase_rotateMasterUserPassword - When true, the master user password is changed to a new strong password generated by Lightsail.

Use the get relational database master user password operation to get the new password.

$sel:disableBackupRetention:UpdateRelationalDatabase', updateRelationalDatabase_disableBackupRetention - When true, disables automated backup retention for your database.

Disabling backup retention deletes all automated database backups. Before disabling this, you may want to create a snapshot of your database using the create relational database snapshot operation.

Updates are applied during the next maintenance window because this can result in an outage.

$sel:relationalDatabaseName:UpdateRelationalDatabase', updateRelationalDatabase_relationalDatabaseName - The name of your Lightsail database resource to update.

Request Lenses

updateRelationalDatabase_masterUserPassword :: Lens' UpdateRelationalDatabase (Maybe Text)

The password for the master user. The password can include any printable ASCII character except "/", """, or "@".

MySQL

Constraints: Must contain from 8 to 41 characters.

PostgreSQL

Constraints: Must contain from 8 to 128 characters.

updateRelationalDatabase_publiclyAccessible :: Lens' UpdateRelationalDatabase (Maybe Bool)

Specifies the accessibility options for your database. A value of true specifies a database that is available to resources outside of your Lightsail account. A value of false specifies a database that is available only to your Lightsail resources in the same region as your database.

updateRelationalDatabase_enableBackupRetention :: Lens' UpdateRelationalDatabase (Maybe Bool)

When true, enables automated backup retention for your database.

Updates are applied during the next maintenance window because this can result in an outage.

updateRelationalDatabase_preferredMaintenanceWindow :: Lens' UpdateRelationalDatabase (Maybe Text)

The weekly time range during which system maintenance can occur on your database.

The default is a 30-minute window selected at random from an 8-hour block of time for each AWS Region, occurring on a random day of the week.

Constraints:

Must be in the ddd:hh24:mi-ddd:hh24:mi format.

Valid days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Must be at least 30 minutes.

Specified in Coordinated Universal Time (UTC).

Example: Tue:17:00-Tue:17:30

updateRelationalDatabase_caCertificateIdentifier :: Lens' UpdateRelationalDatabase (Maybe Text)

Indicates the certificate that needs to be associated with the database.

updateRelationalDatabase_preferredBackupWindow :: Lens' UpdateRelationalDatabase (Maybe Text)

The daily time range during which automated backups are created for your database if automated backups are enabled.

Constraints:

Must be in the hh24:mi-hh24:mi format.

Example: 16:00-16:30

Specified in Coordinated Universal Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

updateRelationalDatabase_applyImmediately :: Lens' UpdateRelationalDatabase (Maybe Bool)

When true, applies changes immediately. When false, applies changes during the preferred maintenance window. Some changes may cause an outage.

Default: false

updateRelationalDatabase_rotateMasterUserPassword :: Lens' UpdateRelationalDatabase (Maybe Bool)

When true, the master user password is changed to a new strong password generated by Lightsail.

Use the get relational database master user password operation to get the new password.

updateRelationalDatabase_disableBackupRetention :: Lens' UpdateRelationalDatabase (Maybe Bool)

When true, disables automated backup retention for your database.

Disabling backup retention deletes all automated database backups. Before disabling this, you may want to create a snapshot of your database using the create relational database snapshot operation.

Updates are applied during the next maintenance window because this can result in an outage.

updateRelationalDatabase_relationalDatabaseName :: Lens' UpdateRelationalDatabase Text

The name of your Lightsail database resource to update.

Destructuring the Response

data UpdateRelationalDatabaseResponse

See: newUpdateRelationalDatabaseResponse smart constructor.

Constructors

UpdateRelationalDatabaseResponse'

Fields

operations :: Maybe [Operation]

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

httpStatus :: Int

The response's http status code.

newUpdateRelationalDatabaseResponse

Arguments

:: Int
$sel:httpStatus:UpdateRelationalDatabaseResponse'

-> UpdateRelationalDatabaseResponse

Create a value of UpdateRelationalDatabaseResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operations:UpdateRelationalDatabaseResponse', updateRelationalDatabaseResponse_operations - An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

$sel:httpStatus:UpdateRelationalDatabaseResponse', updateRelationalDatabaseResponse_httpStatus - The response's http status code.

Response Lenses

updateRelationalDatabaseResponse_operations :: Lens' UpdateRelationalDatabaseResponse (Maybe [Operation])

An array of objects that describe the result of the action, such as the status of the request, the timestamp of the request, and the resources affected by the request.

updateRelationalDatabaseResponse_httpStatus :: Lens' UpdateRelationalDatabaseResponse Int

The response's http status code.