Network.AWS.Location.BatchEvaluateGeofences

Description

Evaluates device positions against the geofence geometries from a given geofence collection.

This operation always returns an empty response because geofences are asynchronously evaluated. The evaluation determines if the device has entered or exited a geofenced area, and then publishes one of the following events to Amazon EventBridge:

ENTER if Amazon Location determines that the tracked device has entered a geofenced area.

EXIT if Amazon Location determines that the tracked device has exited a geofenced area.

The last geofence that a device was observed within is tracked for 30 days after the most recent device position update.

Creating a Request

data BatchEvaluateGeofences

See: newBatchEvaluateGeofences smart constructor.

Constructors

BatchEvaluateGeofences'

Fields

collectionName :: Text

The geofence collection used in evaluating the position of devices against its geofences.

devicePositionUpdates :: NonEmpty DevicePositionUpdate

Contains device details for each device to be evaluated against the given geofence collection.

newBatchEvaluateGeofences

Arguments

:: Text
$sel:collectionName:BatchEvaluateGeofences'

-> NonEmpty DevicePositionUpdate
$sel:devicePositionUpdates:BatchEvaluateGeofences'

-> BatchEvaluateGeofences

Create a value of BatchEvaluateGeofences with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:collectionName:BatchEvaluateGeofences', batchEvaluateGeofences_collectionName - The geofence collection used in evaluating the position of devices against its geofences.

$sel:devicePositionUpdates:BatchEvaluateGeofences', batchEvaluateGeofences_devicePositionUpdates - Contains device details for each device to be evaluated against the given geofence collection.

Request Lenses

batchEvaluateGeofences_collectionName :: Lens' BatchEvaluateGeofences Text

The geofence collection used in evaluating the position of devices against its geofences.

batchEvaluateGeofences_devicePositionUpdates :: Lens' BatchEvaluateGeofences (NonEmpty DevicePositionUpdate)

Contains device details for each device to be evaluated against the given geofence collection.

Destructuring the Response

data BatchEvaluateGeofencesResponse

See: newBatchEvaluateGeofencesResponse smart constructor.

Constructors

BatchEvaluateGeofencesResponse'

Fields

httpStatus :: Int

The response's http status code.

errors :: [BatchEvaluateGeofencesError]

Contains error details for each device that failed to evaluate its position against the given geofence collection.

newBatchEvaluateGeofencesResponse

Arguments

:: Int
$sel:httpStatus:BatchEvaluateGeofencesResponse'

-> BatchEvaluateGeofencesResponse

Create a value of BatchEvaluateGeofencesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:BatchEvaluateGeofencesResponse', batchEvaluateGeofencesResponse_httpStatus - The response's http status code.

$sel:errors:BatchEvaluateGeofencesResponse', batchEvaluateGeofencesResponse_errors - Contains error details for each device that failed to evaluate its position against the given geofence collection.

Response Lenses

batchEvaluateGeofencesResponse_httpStatus :: Lens' BatchEvaluateGeofencesResponse Int

The response's http status code.

batchEvaluateGeofencesResponse_errors :: Lens' BatchEvaluateGeofencesResponse [BatchEvaluateGeofencesError]

Contains error details for each device that failed to evaluate its position against the given geofence collection.