Network.AWS.Location.BatchPutGeofence

Description

A batch request for storing geofence geometries into a given geofence collection, or updates the geometry of an existing geofence if a geofence ID is included in the request.

Creating a Request

data BatchPutGeofence

See: newBatchPutGeofence smart constructor.

Constructors

BatchPutGeofence'

Fields

collectionName :: Text

The geofence collection storing the geofences.

entries :: NonEmpty BatchPutGeofenceRequestEntry

The batch of geofences to be stored in a geofence collection.

newBatchPutGeofence

Arguments

:: Text
$sel:collectionName:BatchPutGeofence'

-> NonEmpty BatchPutGeofenceRequestEntry
$sel:entries:BatchPutGeofence'

-> BatchPutGeofence

Create a value of BatchPutGeofence with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:collectionName:BatchPutGeofence', batchPutGeofence_collectionName - The geofence collection storing the geofences.

$sel:entries:BatchPutGeofence', batchPutGeofence_entries - The batch of geofences to be stored in a geofence collection.

Request Lenses

batchPutGeofence_collectionName :: Lens' BatchPutGeofence Text

The geofence collection storing the geofences.

batchPutGeofence_entries :: Lens' BatchPutGeofence (NonEmpty BatchPutGeofenceRequestEntry)

The batch of geofences to be stored in a geofence collection.

Destructuring the Response

data BatchPutGeofenceResponse

See: newBatchPutGeofenceResponse smart constructor.

Constructors

BatchPutGeofenceResponse'

Fields

httpStatus :: Int

The response's http status code.

errors :: [BatchPutGeofenceError]

Contains additional error details for each geofence that failed to be stored in a geofence collection.

successes :: [BatchPutGeofenceSuccess]

Contains each geofence that was successfully stored in a geofence collection.

newBatchPutGeofenceResponse

Arguments

:: Int
$sel:httpStatus:BatchPutGeofenceResponse'

-> BatchPutGeofenceResponse

Create a value of BatchPutGeofenceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:BatchPutGeofenceResponse', batchPutGeofenceResponse_httpStatus - The response's http status code.

$sel:errors:BatchPutGeofenceResponse', batchPutGeofenceResponse_errors - Contains additional error details for each geofence that failed to be stored in a geofence collection.

$sel:successes:BatchPutGeofenceResponse', batchPutGeofenceResponse_successes - Contains each geofence that was successfully stored in a geofence collection.

Response Lenses

batchPutGeofenceResponse_httpStatus :: Lens' BatchPutGeofenceResponse Int

The response's http status code.

batchPutGeofenceResponse_errors :: Lens' BatchPutGeofenceResponse [BatchPutGeofenceError]

Contains additional error details for each geofence that failed to be stored in a geofence collection.

batchPutGeofenceResponse_successes :: Lens' BatchPutGeofenceResponse [BatchPutGeofenceSuccess]

Contains each geofence that was successfully stored in a geofence collection.