Network.AWS.Location.DeleteMap

Description

Deletes a map resource from your AWS account.

This operation deletes the resource permanently. If the map is being used in an application, the map may not render.

Creating a Request

data DeleteMap

See: newDeleteMap smart constructor.

Constructors

DeleteMap'

Fields

mapName :: Text

The name of the map resource to be deleted.

newDeleteMap

Arguments

:: Text
$sel:mapName:DeleteMap'

-> DeleteMap

Create a value of DeleteMap with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mapName:DeleteMap', deleteMap_mapName - The name of the map resource to be deleted.

Request Lenses

deleteMap_mapName :: Lens' DeleteMap Text

The name of the map resource to be deleted.

Destructuring the Response

data DeleteMapResponse

See: newDeleteMapResponse smart constructor.

Constructors

DeleteMapResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteMapResponse

Arguments

:: Int
$sel:httpStatus:DeleteMapResponse'

-> DeleteMapResponse

Create a value of DeleteMapResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteMapResponse', deleteMapResponse_httpStatus - The response's http status code.

Response Lenses

deleteMapResponse_httpStatus :: Lens' DeleteMapResponse Int

The response's http status code.