Network.AWS.Location.DeleteRouteCalculator

Description

Deletes a route calculator resource from your AWS account.

This operation deletes the resource permanently.

Creating a Request

data DeleteRouteCalculator

See: newDeleteRouteCalculator smart constructor.

Constructors

DeleteRouteCalculator'

Fields

calculatorName :: Text

The name of the route calculator resource to be deleted.

newDeleteRouteCalculator

Arguments

:: Text
$sel:calculatorName:DeleteRouteCalculator'

-> DeleteRouteCalculator

Create a value of DeleteRouteCalculator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:calculatorName:DeleteRouteCalculator', deleteRouteCalculator_calculatorName - The name of the route calculator resource to be deleted.

Request Lenses

deleteRouteCalculator_calculatorName :: Lens' DeleteRouteCalculator Text

The name of the route calculator resource to be deleted.

Destructuring the Response

data DeleteRouteCalculatorResponse

See: newDeleteRouteCalculatorResponse smart constructor.

Constructors

DeleteRouteCalculatorResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteRouteCalculatorResponse

Arguments

:: Int
$sel:httpStatus:DeleteRouteCalculatorResponse'

-> DeleteRouteCalculatorResponse

Create a value of DeleteRouteCalculatorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteRouteCalculatorResponse', deleteRouteCalculatorResponse_httpStatus - The response's http status code.

Response Lenses

deleteRouteCalculatorResponse_httpStatus :: Lens' DeleteRouteCalculatorResponse Int

The response's http status code.