Network.AWS.Location.DescribeGeofenceCollection

Description

Retrieves the geofence collection details.

Creating a Request

data DescribeGeofenceCollection

See: newDescribeGeofenceCollection smart constructor.

Constructors

DescribeGeofenceCollection'

Fields

collectionName :: Text

The name of the geofence collection.

newDescribeGeofenceCollection

Arguments

:: Text
$sel:collectionName:DescribeGeofenceCollection'

-> DescribeGeofenceCollection

Create a value of DescribeGeofenceCollection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:collectionName:DescribeGeofenceCollection', describeGeofenceCollection_collectionName - The name of the geofence collection.

Request Lenses

describeGeofenceCollection_collectionName :: Lens' DescribeGeofenceCollection Text

The name of the geofence collection.

Destructuring the Response

data DescribeGeofenceCollectionResponse

See: newDescribeGeofenceCollectionResponse smart constructor.

Constructors

DescribeGeofenceCollectionResponse'

Fields

pricingPlanDataSource :: Maybe Text

The specified data provider for the geofence collection.

kmsKeyId :: Maybe Text

A key identifier for an AWS KMS customer managed key assigned to the Amazon Location resource

tags :: Maybe (HashMap Text Text)

Displays the key, value pairs of tags associated with this resource.

httpStatus :: Int

The response's http status code.

collectionArn :: Text

The Amazon Resource Name (ARN) for the geofence collection resource. Used when you need to specify a resource across all AWS.

Format example: arn:aws:geo:region:account-id:geofence-collection/ExampleGeofenceCollection

collectionName :: Text

The name of the geofence collection.

createTime :: POSIX

The timestamp for when the geofence resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

description :: Text

The optional description for the geofence collection.

pricingPlan :: PricingPlan

The pricing plan selected for the specified geofence collection.

For additional details and restrictions on each pricing plan option, see the Amazon Location Service pricing page.

updateTime :: POSIX

The timestamp for when the geofence collection was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

newDescribeGeofenceCollectionResponse

Arguments

:: Int
$sel:httpStatus:DescribeGeofenceCollectionResponse'

-> Text
$sel:collectionArn:DescribeGeofenceCollectionResponse'

-> Text
$sel:collectionName:DescribeGeofenceCollection'

-> UTCTime
$sel:createTime:DescribeGeofenceCollectionResponse'

-> Text
$sel:description:DescribeGeofenceCollectionResponse'

-> PricingPlan
$sel:pricingPlan:DescribeGeofenceCollectionResponse'

-> UTCTime
$sel:updateTime:DescribeGeofenceCollectionResponse'

-> DescribeGeofenceCollectionResponse

Create a value of DescribeGeofenceCollectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pricingPlanDataSource:DescribeGeofenceCollectionResponse', describeGeofenceCollectionResponse_pricingPlanDataSource - The specified data provider for the geofence collection.

$sel:kmsKeyId:DescribeGeofenceCollectionResponse', describeGeofenceCollectionResponse_kmsKeyId - A key identifier for an AWS KMS customer managed key assigned to the Amazon Location resource

$sel:tags:DescribeGeofenceCollectionResponse', describeGeofenceCollectionResponse_tags - Displays the key, value pairs of tags associated with this resource.

$sel:httpStatus:DescribeGeofenceCollectionResponse', describeGeofenceCollectionResponse_httpStatus - The response's http status code.

$sel:collectionArn:DescribeGeofenceCollectionResponse', describeGeofenceCollectionResponse_collectionArn - The Amazon Resource Name (ARN) for the geofence collection resource. Used when you need to specify a resource across all AWS.

Format example: arn:aws:geo:region:account-id:geofence-collection/ExampleGeofenceCollection

$sel:collectionName:DescribeGeofenceCollection', describeGeofenceCollectionResponse_collectionName - The name of the geofence collection.

$sel:createTime:DescribeGeofenceCollectionResponse', describeGeofenceCollectionResponse_createTime - The timestamp for when the geofence resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

$sel:description:DescribeGeofenceCollectionResponse', describeGeofenceCollectionResponse_description - The optional description for the geofence collection.

$sel:pricingPlan:DescribeGeofenceCollectionResponse', describeGeofenceCollectionResponse_pricingPlan - The pricing plan selected for the specified geofence collection.

For additional details and restrictions on each pricing plan option, see the Amazon Location Service pricing page.

$sel:updateTime:DescribeGeofenceCollectionResponse', describeGeofenceCollectionResponse_updateTime - The timestamp for when the geofence collection was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

Response Lenses

describeGeofenceCollectionResponse_pricingPlanDataSource :: Lens' DescribeGeofenceCollectionResponse (Maybe Text)

The specified data provider for the geofence collection.

describeGeofenceCollectionResponse_kmsKeyId :: Lens' DescribeGeofenceCollectionResponse (Maybe Text)

A key identifier for an AWS KMS customer managed key assigned to the Amazon Location resource

describeGeofenceCollectionResponse_tags :: Lens' DescribeGeofenceCollectionResponse (Maybe (HashMap Text Text))

Displays the key, value pairs of tags associated with this resource.

describeGeofenceCollectionResponse_httpStatus :: Lens' DescribeGeofenceCollectionResponse Int

The response's http status code.

describeGeofenceCollectionResponse_collectionArn :: Lens' DescribeGeofenceCollectionResponse Text

The Amazon Resource Name (ARN) for the geofence collection resource. Used when you need to specify a resource across all AWS.

Format example: arn:aws:geo:region:account-id:geofence-collection/ExampleGeofenceCollection

describeGeofenceCollectionResponse_collectionName :: Lens' DescribeGeofenceCollectionResponse Text

The name of the geofence collection.

describeGeofenceCollectionResponse_createTime :: Lens' DescribeGeofenceCollectionResponse UTCTime

The timestamp for when the geofence resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

describeGeofenceCollectionResponse_description :: Lens' DescribeGeofenceCollectionResponse Text

The optional description for the geofence collection.

describeGeofenceCollectionResponse_pricingPlan :: Lens' DescribeGeofenceCollectionResponse PricingPlan

The pricing plan selected for the specified geofence collection.

For additional details and restrictions on each pricing plan option, see the Amazon Location Service pricing page.

describeGeofenceCollectionResponse_updateTime :: Lens' DescribeGeofenceCollectionResponse UTCTime

The timestamp for when the geofence collection was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ