Network.AWS.Location.DescribeMap

Description

Retrieves the map resource details.

Creating a Request

data DescribeMap

See: newDescribeMap smart constructor.

Constructors

DescribeMap'

Fields

mapName :: Text

The name of the map resource.

newDescribeMap

Arguments

:: Text
$sel:mapName:DescribeMap'

-> DescribeMap

Create a value of DescribeMap with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mapName:DescribeMap', describeMap_mapName - The name of the map resource.

Request Lenses

describeMap_mapName :: Lens' DescribeMap Text

The name of the map resource.

Destructuring the Response

data DescribeMapResponse

See: newDescribeMapResponse smart constructor.

Constructors

DescribeMapResponse'

Fields

tags :: Maybe (HashMap Text Text)

Tags associated with the map resource.

httpStatus :: Int

The response's http status code.

configuration :: MapConfiguration

Specifies the map tile style selected from a partner data provider.

createTime :: POSIX

The timestamp for when the map resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

dataSource :: Text

Specifies the data provider for the associated map tiles.

description :: Text

The optional description for the map resource.

mapArn :: Text

The Amazon Resource Name (ARN) for the map resource. Used to specify a resource across all AWS.

Format example: arn:aws:geo:region:account-id:maps/ExampleMap

mapName :: Text

The map style selected from an available provider.

pricingPlan :: PricingPlan

The pricing plan selected for the specified map resource.

 <p>For additional details and restrictions on each pricing plan option, see <a href="https://aws.amazon.com/location/pricing/">Amazon Location Service pricing</a>.</p>

updateTime :: POSIX

The timestamp for when the map resource was last update in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

newDescribeMapResponse

Arguments

:: Int
$sel:httpStatus:DescribeMapResponse'

-> MapConfiguration
$sel:configuration:DescribeMapResponse'

-> UTCTime
$sel:createTime:DescribeMapResponse'

-> Text
$sel:dataSource:DescribeMapResponse'

-> Text
$sel:description:DescribeMapResponse'

-> Text
$sel:mapArn:DescribeMapResponse'

-> Text
$sel:mapName:DescribeMap'

-> PricingPlan
$sel:pricingPlan:DescribeMapResponse'

-> UTCTime
$sel:updateTime:DescribeMapResponse'

-> DescribeMapResponse

Create a value of DescribeMapResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:DescribeMapResponse', describeMapResponse_tags - Tags associated with the map resource.

$sel:httpStatus:DescribeMapResponse', describeMapResponse_httpStatus - The response's http status code.

$sel:configuration:DescribeMapResponse', describeMapResponse_configuration - Specifies the map tile style selected from a partner data provider.

$sel:createTime:DescribeMapResponse', describeMapResponse_createTime - The timestamp for when the map resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

$sel:dataSource:DescribeMapResponse', describeMapResponse_dataSource - Specifies the data provider for the associated map tiles.

$sel:description:DescribeMapResponse', describeMapResponse_description - The optional description for the map resource.

$sel:mapArn:DescribeMapResponse', describeMapResponse_mapArn - The Amazon Resource Name (ARN) for the map resource. Used to specify a resource across all AWS.

Format example: arn:aws:geo:region:account-id:maps/ExampleMap

$sel:mapName:DescribeMap', describeMapResponse_mapName - The map style selected from an available provider.

$sel:pricingPlan:DescribeMapResponse', describeMapResponse_pricingPlan - The pricing plan selected for the specified map resource.

 <p>For additional details and restrictions on each pricing plan option, see <a href="https://aws.amazon.com/location/pricing/">Amazon Location Service pricing</a>.</p>

$sel:updateTime:DescribeMapResponse', describeMapResponse_updateTime - The timestamp for when the map resource was last update in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

Response Lenses

describeMapResponse_tags :: Lens' DescribeMapResponse (Maybe (HashMap Text Text))

Tags associated with the map resource.

describeMapResponse_httpStatus :: Lens' DescribeMapResponse Int

The response's http status code.

describeMapResponse_configuration :: Lens' DescribeMapResponse MapConfiguration

Specifies the map tile style selected from a partner data provider.

describeMapResponse_createTime :: Lens' DescribeMapResponse UTCTime

The timestamp for when the map resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

describeMapResponse_dataSource :: Lens' DescribeMapResponse Text

Specifies the data provider for the associated map tiles.

describeMapResponse_description :: Lens' DescribeMapResponse Text

The optional description for the map resource.

describeMapResponse_mapArn :: Lens' DescribeMapResponse Text

The Amazon Resource Name (ARN) for the map resource. Used to specify a resource across all AWS.

Format example: arn:aws:geo:region:account-id:maps/ExampleMap

describeMapResponse_mapName :: Lens' DescribeMapResponse Text

The map style selected from an available provider.

describeMapResponse_pricingPlan :: Lens' DescribeMapResponse PricingPlan

The pricing plan selected for the specified map resource.

 <p>For additional details and restrictions on each pricing plan option, see <a href="https://aws.amazon.com/location/pricing/">Amazon Location Service pricing</a>.</p>

describeMapResponse_updateTime :: Lens' DescribeMapResponse UTCTime

The timestamp for when the map resource was last update in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.