Network.AWS.Location.DescribeRouteCalculator

Description

Retrieves the route calculator resource details.

Creating a Request

data DescribeRouteCalculator

See: newDescribeRouteCalculator smart constructor.

Constructors

DescribeRouteCalculator'

Fields

calculatorName :: Text

The name of the route calculator resource.

newDescribeRouteCalculator

Arguments

:: Text
$sel:calculatorName:DescribeRouteCalculator'

-> DescribeRouteCalculator

Create a value of DescribeRouteCalculator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:calculatorName:DescribeRouteCalculator', describeRouteCalculator_calculatorName - The name of the route calculator resource.

Request Lenses

describeRouteCalculator_calculatorName :: Lens' DescribeRouteCalculator Text

The name of the route calculator resource.

Destructuring the Response

data DescribeRouteCalculatorResponse

See: newDescribeRouteCalculatorResponse smart constructor.

Constructors

DescribeRouteCalculatorResponse'

Fields

tags :: Maybe (HashMap Text Text)

Tags associated with route calculator resource.

httpStatus :: Int

The response's http status code.

calculatorArn :: Text

The Amazon Resource Name (ARN) for the Route calculator resource. Use the ARN when you specify a resource across AWS.

Format example: arn:aws:geo:region:account-id:route-calculator/ExampleCalculator

calculatorName :: Text

The name of the route calculator resource being described.

createTime :: POSIX

The timestamp when the route calculator resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

For example, 2020–07-2T12:15:20.000Z+01:00

dataSource :: Text

The data provider of traffic and road network data. Indicates one of the available providers:

Esri

Here

For more information about data providers, see Amazon Location Service data providers.

description :: Text

The optional description of the route calculator resource.

pricingPlan :: PricingPlan

The pricing plan selected for the specified route calculator resource.

For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

updateTime :: POSIX

The timestamp when the route calculator resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

For example, 2020–07-2T12:15:20.000Z+01:00

newDescribeRouteCalculatorResponse

Arguments

:: Int
$sel:httpStatus:DescribeRouteCalculatorResponse'

-> Text
$sel:calculatorArn:DescribeRouteCalculatorResponse'

-> Text
$sel:calculatorName:DescribeRouteCalculator'

-> UTCTime
$sel:createTime:DescribeRouteCalculatorResponse'

-> Text
$sel:dataSource:DescribeRouteCalculatorResponse'

-> Text
$sel:description:DescribeRouteCalculatorResponse'

-> PricingPlan
$sel:pricingPlan:DescribeRouteCalculatorResponse'

-> UTCTime
$sel:updateTime:DescribeRouteCalculatorResponse'

-> DescribeRouteCalculatorResponse

Create a value of DescribeRouteCalculatorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:DescribeRouteCalculatorResponse', describeRouteCalculatorResponse_tags - Tags associated with route calculator resource.

$sel:httpStatus:DescribeRouteCalculatorResponse', describeRouteCalculatorResponse_httpStatus - The response's http status code.

$sel:calculatorArn:DescribeRouteCalculatorResponse', describeRouteCalculatorResponse_calculatorArn - The Amazon Resource Name (ARN) for the Route calculator resource. Use the ARN when you specify a resource across AWS.

Format example: arn:aws:geo:region:account-id:route-calculator/ExampleCalculator

$sel:calculatorName:DescribeRouteCalculator', describeRouteCalculatorResponse_calculatorName - The name of the route calculator resource being described.

$sel:createTime:DescribeRouteCalculatorResponse', describeRouteCalculatorResponse_createTime - The timestamp when the route calculator resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

For example, 2020–07-2T12:15:20.000Z+01:00

$sel:dataSource:DescribeRouteCalculatorResponse', describeRouteCalculatorResponse_dataSource - The data provider of traffic and road network data. Indicates one of the available providers:

Esri

Here

For more information about data providers, see Amazon Location Service data providers.

$sel:description:DescribeRouteCalculatorResponse', describeRouteCalculatorResponse_description - The optional description of the route calculator resource.

$sel:pricingPlan:DescribeRouteCalculatorResponse', describeRouteCalculatorResponse_pricingPlan - The pricing plan selected for the specified route calculator resource.

For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

$sel:updateTime:DescribeRouteCalculatorResponse', describeRouteCalculatorResponse_updateTime - The timestamp when the route calculator resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

For example, 2020–07-2T12:15:20.000Z+01:00

Response Lenses

describeRouteCalculatorResponse_tags :: Lens' DescribeRouteCalculatorResponse (Maybe (HashMap Text Text))

Tags associated with route calculator resource.

describeRouteCalculatorResponse_httpStatus :: Lens' DescribeRouteCalculatorResponse Int

The response's http status code.

describeRouteCalculatorResponse_calculatorArn :: Lens' DescribeRouteCalculatorResponse Text

The Amazon Resource Name (ARN) for the Route calculator resource. Use the ARN when you specify a resource across AWS.

Format example: arn:aws:geo:region:account-id:route-calculator/ExampleCalculator

describeRouteCalculatorResponse_calculatorName :: Lens' DescribeRouteCalculatorResponse Text

The name of the route calculator resource being described.

describeRouteCalculatorResponse_createTime :: Lens' DescribeRouteCalculatorResponse UTCTime

The timestamp when the route calculator resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

For example, 2020–07-2T12:15:20.000Z+01:00

describeRouteCalculatorResponse_dataSource :: Lens' DescribeRouteCalculatorResponse Text

The data provider of traffic and road network data. Indicates one of the available providers:

Esri

Here

For more information about data providers, see Amazon Location Service data providers.

describeRouteCalculatorResponse_description :: Lens' DescribeRouteCalculatorResponse Text

The optional description of the route calculator resource.

describeRouteCalculatorResponse_pricingPlan :: Lens' DescribeRouteCalculatorResponse PricingPlan

The pricing plan selected for the specified route calculator resource.

For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

describeRouteCalculatorResponse_updateTime :: Lens' DescribeRouteCalculatorResponse UTCTime

The timestamp when the route calculator resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

For example, 2020–07-2T12:15:20.000Z+01:00