Network.AWS.Location.DescribeTracker

Description

Retrieves the tracker resource details.

Creating a Request

data DescribeTracker

See: newDescribeTracker smart constructor.

Constructors

DescribeTracker'

Fields

trackerName :: Text

The name of the tracker resource.

newDescribeTracker

Arguments

:: Text
$sel:trackerName:DescribeTracker'

-> DescribeTracker

Create a value of DescribeTracker with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trackerName:DescribeTracker', describeTracker_trackerName - The name of the tracker resource.

Request Lenses

describeTracker_trackerName :: Lens' DescribeTracker Text

The name of the tracker resource.

Destructuring the Response

data DescribeTrackerResponse

See: newDescribeTrackerResponse smart constructor.

Constructors

DescribeTrackerResponse'

Fields

pricingPlanDataSource :: Maybe Text

The specified data provider for the tracker resource.

kmsKeyId :: Maybe Text

A key identifier for an AWS KMS customer managed key assigned to the Amazon Location resource.

tags :: Maybe (HashMap Text Text)

The tags associated with the tracker resource.

positionFiltering :: Maybe PositionFiltering

The position filtering method of the tracker resource.

httpStatus :: Int

The response's http status code.

createTime :: POSIX

The timestamp for when the tracker resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

description :: Text

The optional description for the tracker resource.

pricingPlan :: PricingPlan

The pricing plan selected for the specified tracker resource.

For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

trackerArn :: Text

The Amazon Resource Name (ARN) for the tracker resource. Used when you need to specify a resource across all AWS.

Format example: arn:aws:geo:region:account-id:tracker/ExampleTracker

trackerName :: Text

The name of the tracker resource.

updateTime :: POSIX

The timestamp for when the tracker resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

newDescribeTrackerResponse

Arguments

:: Int
$sel:httpStatus:DescribeTrackerResponse'

-> UTCTime
$sel:createTime:DescribeTrackerResponse'

-> Text
$sel:description:DescribeTrackerResponse'

-> PricingPlan
$sel:pricingPlan:DescribeTrackerResponse'

-> Text
$sel:trackerArn:DescribeTrackerResponse'

-> Text
$sel:trackerName:DescribeTracker'

-> UTCTime
$sel:updateTime:DescribeTrackerResponse'

-> DescribeTrackerResponse

Create a value of DescribeTrackerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pricingPlanDataSource:DescribeTrackerResponse', describeTrackerResponse_pricingPlanDataSource - The specified data provider for the tracker resource.

$sel:kmsKeyId:DescribeTrackerResponse', describeTrackerResponse_kmsKeyId - A key identifier for an AWS KMS customer managed key assigned to the Amazon Location resource.

$sel:tags:DescribeTrackerResponse', describeTrackerResponse_tags - The tags associated with the tracker resource.

$sel:positionFiltering:DescribeTrackerResponse', describeTrackerResponse_positionFiltering - The position filtering method of the tracker resource.

$sel:httpStatus:DescribeTrackerResponse', describeTrackerResponse_httpStatus - The response's http status code.

$sel:createTime:DescribeTrackerResponse', describeTrackerResponse_createTime - The timestamp for when the tracker resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

$sel:description:DescribeTrackerResponse', describeTrackerResponse_description - The optional description for the tracker resource.

$sel:pricingPlan:DescribeTrackerResponse', describeTrackerResponse_pricingPlan - The pricing plan selected for the specified tracker resource.

For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

$sel:trackerArn:DescribeTrackerResponse', describeTrackerResponse_trackerArn - The Amazon Resource Name (ARN) for the tracker resource. Used when you need to specify a resource across all AWS.

Format example: arn:aws:geo:region:account-id:tracker/ExampleTracker

$sel:trackerName:DescribeTracker', describeTrackerResponse_trackerName - The name of the tracker resource.

$sel:updateTime:DescribeTrackerResponse', describeTrackerResponse_updateTime - The timestamp for when the tracker resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

Response Lenses

describeTrackerResponse_pricingPlanDataSource :: Lens' DescribeTrackerResponse (Maybe Text)

The specified data provider for the tracker resource.

describeTrackerResponse_kmsKeyId :: Lens' DescribeTrackerResponse (Maybe Text)

A key identifier for an AWS KMS customer managed key assigned to the Amazon Location resource.

describeTrackerResponse_tags :: Lens' DescribeTrackerResponse (Maybe (HashMap Text Text))

The tags associated with the tracker resource.

describeTrackerResponse_positionFiltering :: Lens' DescribeTrackerResponse (Maybe PositionFiltering)

The position filtering method of the tracker resource.

describeTrackerResponse_httpStatus :: Lens' DescribeTrackerResponse Int

The response's http status code.

describeTrackerResponse_createTime :: Lens' DescribeTrackerResponse UTCTime

The timestamp for when the tracker resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

describeTrackerResponse_description :: Lens' DescribeTrackerResponse Text

The optional description for the tracker resource.

describeTrackerResponse_pricingPlan :: Lens' DescribeTrackerResponse PricingPlan

The pricing plan selected for the specified tracker resource.

For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

describeTrackerResponse_trackerArn :: Lens' DescribeTrackerResponse Text

The Amazon Resource Name (ARN) for the tracker resource. Used when you need to specify a resource across all AWS.

Format example: arn:aws:geo:region:account-id:tracker/ExampleTracker

describeTrackerResponse_trackerName :: Lens' DescribeTrackerResponse Text

The name of the tracker resource.

describeTrackerResponse_updateTime :: Lens' DescribeTrackerResponse UTCTime

The timestamp for when the tracker resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.