Network.AWS.Location.GetMapSprites

Description

Retrieves the sprite sheet corresponding to a map resource. The sprite sheet is a PNG image paired with a JSON document describing the offsets of individual icons that will be displayed on a rendered map.

Creating a Request

data GetMapSprites

See: newGetMapSprites smart constructor.

Constructors

GetMapSprites'

Fields

fileName :: Text

The name of the sprite ﬁle. Use the following ﬁle names for the sprite sheet:

sprites.png

sprites@2x.png for high pixel density displays

For the JSON document contain image offsets. Use the following ﬁle names:

sprites.json

sprites@2x.json for high pixel density displays

mapName :: Text

The map resource associated with the sprite ﬁle.

newGetMapSprites

Arguments

:: Text
$sel:fileName:GetMapSprites'

-> Text
$sel:mapName:GetMapSprites'

-> GetMapSprites

Create a value of GetMapSprites with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fileName:GetMapSprites', getMapSprites_fileName - The name of the sprite ﬁle. Use the following ﬁle names for the sprite sheet:

sprites.png

sprites@2x.png for high pixel density displays

For the JSON document contain image offsets. Use the following ﬁle names:

sprites.json

sprites@2x.json for high pixel density displays

$sel:mapName:GetMapSprites', getMapSprites_mapName - The map resource associated with the sprite ﬁle.

Request Lenses

getMapSprites_fileName :: Lens' GetMapSprites Text

The name of the sprite ﬁle. Use the following ﬁle names for the sprite sheet:

sprites.png

sprites@2x.png for high pixel density displays

For the JSON document contain image offsets. Use the following ﬁle names:

sprites.json

sprites@2x.json for high pixel density displays

getMapSprites_mapName :: Lens' GetMapSprites Text

The map resource associated with the sprite ﬁle.

Destructuring the Response

data GetMapSpritesResponse

See: newGetMapSpritesResponse smart constructor.

Constructors

GetMapSpritesResponse'

Fields

blob :: Maybe ByteString

Contains the body of the sprite sheet or JSON offset ﬁle.

contentType :: Maybe Text

The content type of the sprite sheet and offsets. For example, the sprite sheet content type is image/png, and the sprite offset JSON document is application/json.

httpStatus :: Int

The response's http status code.

newGetMapSpritesResponse

Arguments

:: Int
$sel:httpStatus:GetMapSpritesResponse'

-> GetMapSpritesResponse

Create a value of GetMapSpritesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blob:GetMapSpritesResponse', getMapSpritesResponse_blob - Contains the body of the sprite sheet or JSON offset ﬁle.

$sel:contentType:GetMapSpritesResponse', getMapSpritesResponse_contentType - The content type of the sprite sheet and offsets. For example, the sprite sheet content type is image/png, and the sprite offset JSON document is application/json.

$sel:httpStatus:GetMapSpritesResponse', getMapSpritesResponse_httpStatus - The response's http status code.

Response Lenses

getMapSpritesResponse_blob :: Lens' GetMapSpritesResponse (Maybe ByteString)

Contains the body of the sprite sheet or JSON offset ﬁle.

getMapSpritesResponse_contentType :: Lens' GetMapSpritesResponse (Maybe Text)

The content type of the sprite sheet and offsets. For example, the sprite sheet content type is image/png, and the sprite offset JSON document is application/json.

getMapSpritesResponse_httpStatus :: Lens' GetMapSpritesResponse Int

The response's http status code.