Network.AWS.Location.GetMapStyleDescriptor

Description

Retrieves the map style descriptor from a map resource.

The style descriptor contains speciﬁcations on how features render on a map. For example, what data to display, what order to display the data in, and the style for the data. Style descriptors follow the Mapbox Style Specification.

Creating a Request

data GetMapStyleDescriptor

See: newGetMapStyleDescriptor smart constructor.

Constructors

GetMapStyleDescriptor'

Fields

mapName :: Text

The map resource to retrieve the style descriptor from.

newGetMapStyleDescriptor

Arguments

:: Text
$sel:mapName:GetMapStyleDescriptor'

-> GetMapStyleDescriptor

Create a value of GetMapStyleDescriptor with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mapName:GetMapStyleDescriptor', getMapStyleDescriptor_mapName - The map resource to retrieve the style descriptor from.

Request Lenses

getMapStyleDescriptor_mapName :: Lens' GetMapStyleDescriptor Text

The map resource to retrieve the style descriptor from.

Destructuring the Response

data GetMapStyleDescriptorResponse

See: newGetMapStyleDescriptorResponse smart constructor.

Constructors

GetMapStyleDescriptorResponse'

Fields

blob :: Maybe ByteString

Contains the body of the style descriptor.

contentType :: Maybe Text

The style descriptor's content type. For example, application/json.

httpStatus :: Int

The response's http status code.

newGetMapStyleDescriptorResponse

Arguments

:: Int
$sel:httpStatus:GetMapStyleDescriptorResponse'

-> GetMapStyleDescriptorResponse

Create a value of GetMapStyleDescriptorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blob:GetMapStyleDescriptorResponse', getMapStyleDescriptorResponse_blob - Contains the body of the style descriptor.

$sel:contentType:GetMapStyleDescriptorResponse', getMapStyleDescriptorResponse_contentType - The style descriptor's content type. For example, application/json.

$sel:httpStatus:GetMapStyleDescriptorResponse', getMapStyleDescriptorResponse_httpStatus - The response's http status code.

Response Lenses

getMapStyleDescriptorResponse_blob :: Lens' GetMapStyleDescriptorResponse (Maybe ByteString)

Contains the body of the style descriptor.

getMapStyleDescriptorResponse_contentType :: Lens' GetMapStyleDescriptorResponse (Maybe Text)

The style descriptor's content type. For example, application/json.

getMapStyleDescriptorResponse_httpStatus :: Lens' GetMapStyleDescriptorResponse Int

The response's http status code.