Network.AWS.Location.GetMapTile

Description

Retrieves a vector data tile from the map resource. Map tiles are used by clients to render a map. they're addressed using a grid arrangement with an X coordinate, Y coordinate, and Z (zoom) level.

The origin (0, 0) is the top left of the map. Increasing the zoom level by 1 doubles both the X and Y dimensions, so a tile containing data for the entire world at (0/0/0) will be split into 4 tiles at zoom 1 (1/0/0, 1/0/1, 1/1/0, 1/1/1).

Creating a Request

data GetMapTile

See: newGetMapTile smart constructor.

Constructors

GetMapTile'

Fields

mapName :: Text

The map resource to retrieve the map tiles from.

x :: Text

The X axis value for the map tile.

y :: Text

The Y axis value for the map tile.

z :: Text

The zoom value for the map tile.

newGetMapTile

Arguments

:: Text
$sel:mapName:GetMapTile'

-> Text
$sel:x:GetMapTile'

-> Text
$sel:y:GetMapTile'

-> Text
$sel:z:GetMapTile'

-> GetMapTile

Create a value of GetMapTile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mapName:GetMapTile', getMapTile_mapName - The map resource to retrieve the map tiles from.

$sel:x:GetMapTile', getMapTile_x - The X axis value for the map tile.

$sel:y:GetMapTile', getMapTile_y - The Y axis value for the map tile.

$sel:z:GetMapTile', getMapTile_z - The zoom value for the map tile.

Request Lenses

getMapTile_mapName :: Lens' GetMapTile Text

The map resource to retrieve the map tiles from.

getMapTile_x :: Lens' GetMapTile Text

The X axis value for the map tile.

getMapTile_y :: Lens' GetMapTile Text

The Y axis value for the map tile.

getMapTile_z :: Lens' GetMapTile Text

The zoom value for the map tile.

Destructuring the Response

data GetMapTileResponse

See: newGetMapTileResponse smart constructor.

Constructors

GetMapTileResponse'

Fields

blob :: Maybe ByteString

Contains Mapbox Vector Tile (MVT) data.

contentType :: Maybe Text

The map tile's content type. For example, application/vnd.mapbox-vector-tile.

httpStatus :: Int

The response's http status code.

newGetMapTileResponse

Arguments

:: Int
$sel:httpStatus:GetMapTileResponse'

-> GetMapTileResponse

Create a value of GetMapTileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blob:GetMapTileResponse', getMapTileResponse_blob - Contains Mapbox Vector Tile (MVT) data.

$sel:contentType:GetMapTileResponse', getMapTileResponse_contentType - The map tile's content type. For example, application/vnd.mapbox-vector-tile.

$sel:httpStatus:GetMapTileResponse', getMapTileResponse_httpStatus - The response's http status code.

Response Lenses

getMapTileResponse_blob :: Lens' GetMapTileResponse (Maybe ByteString)

Contains Mapbox Vector Tile (MVT) data.

getMapTileResponse_contentType :: Lens' GetMapTileResponse (Maybe Text)

The map tile's content type. For example, application/vnd.mapbox-vector-tile.

getMapTileResponse_httpStatus :: Lens' GetMapTileResponse Int

The response's http status code.