Network.AWS.Location.PutGeofence

Description

Stores a geofence geometry in a given geofence collection, or updates the geometry of an existing geofence if a geofence ID is included in the request.

Creating a Request

data PutGeofence

See: newPutGeofence smart constructor.

Constructors

PutGeofence'

Fields

collectionName :: Text

The geofence collection to store the geofence in.

geofenceId :: Text

An identifier for the geofence. For example, ExampleGeofence-1.

geometry :: GeofenceGeometry

Contains the polygon details to specify the position of the geofence.

Each geofence polygon can have a maximum of 1,000 vertices.

newPutGeofence

Arguments

:: Text
$sel:collectionName:PutGeofence'

-> Text
$sel:geofenceId:PutGeofence'

-> GeofenceGeometry
$sel:geometry:PutGeofence'

-> PutGeofence

Create a value of PutGeofence with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:collectionName:PutGeofence', putGeofence_collectionName - The geofence collection to store the geofence in.

$sel:geofenceId:PutGeofence', putGeofence_geofenceId - An identifier for the geofence. For example, ExampleGeofence-1.

$sel:geometry:PutGeofence', putGeofence_geometry - Contains the polygon details to specify the position of the geofence.

Each geofence polygon can have a maximum of 1,000 vertices.

Request Lenses

putGeofence_collectionName :: Lens' PutGeofence Text

The geofence collection to store the geofence in.

putGeofence_geofenceId :: Lens' PutGeofence Text

An identifier for the geofence. For example, ExampleGeofence-1.

putGeofence_geometry :: Lens' PutGeofence GeofenceGeometry

Contains the polygon details to specify the position of the geofence.

Each geofence polygon can have a maximum of 1,000 vertices.

Destructuring the Response

data PutGeofenceResponse

See: newPutGeofenceResponse smart constructor.

Constructors

PutGeofenceResponse'

Fields

httpStatus :: Int

The response's http status code.

createTime :: POSIX

The timestamp for when the geofence was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

geofenceId :: Text

The geofence identifier entered in the request.

updateTime :: POSIX

The timestamp for when the geofence was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

newPutGeofenceResponse

Arguments

:: Int
$sel:httpStatus:PutGeofenceResponse'

-> UTCTime
$sel:createTime:PutGeofenceResponse'

-> Text
$sel:geofenceId:PutGeofence'

-> UTCTime
$sel:updateTime:PutGeofenceResponse'

-> PutGeofenceResponse

Create a value of PutGeofenceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutGeofenceResponse', putGeofenceResponse_httpStatus - The response's http status code.

$sel:createTime:PutGeofenceResponse', putGeofenceResponse_createTime - The timestamp for when the geofence was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

$sel:geofenceId:PutGeofence', putGeofenceResponse_geofenceId - The geofence identifier entered in the request.

$sel:updateTime:PutGeofenceResponse', putGeofenceResponse_updateTime - The timestamp for when the geofence was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

Response Lenses

putGeofenceResponse_httpStatus :: Lens' PutGeofenceResponse Int

The response's http status code.

putGeofenceResponse_createTime :: Lens' PutGeofenceResponse UTCTime

The timestamp for when the geofence was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

putGeofenceResponse_geofenceId :: Lens' PutGeofenceResponse Text

The geofence identifier entered in the request.

putGeofenceResponse_updateTime :: Lens' PutGeofenceResponse UTCTime

The timestamp for when the geofence was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ