Network.AWS.Location.SearchPlaceIndexForText

Description

Geocodes free-form text, such as an address, name, city, or region to allow you to search for Places or points of interest.

Includes the option to apply additional parameters to narrow your list of results.

You can search for places near a given position using BiasPosition, or filter results within a bounding box using FilterBBox. Providing both parameters simultaneously returns an error.

Creating a Request

data SearchPlaceIndexForText

See: newSearchPlaceIndexForText smart constructor.

Constructors

SearchPlaceIndexForText'

Fields

filterBBox :: Maybe (Sensitive (NonEmpty Double))

Filters the results by returning only Places within the provided bounding box. An optional parameter.

The first 2 bbox parameters describe the lower southwest corner:

The first bbox position is the X coordinate or longitude of the lower southwest corner.

The second bbox position is the Y coordinate or latitude of the lower southwest corner.

For example, bbox=xLongitudeSW&bbox=yLatitudeSW.

The next bbox parameters describe the upper northeast corner:

The third bbox position is the X coordinate, or longitude of the upper northeast corner.

The fourth bbox position is the Y coordinate, or longitude of the upper northeast corner.

For example, bbox=xLongitudeNE&bbox=yLatitudeNE

biasPosition :: Maybe (Sensitive (NonEmpty Double))

Searches for results closest to the given position. An optional parameter defined by longitude, and latitude.

The first bias position is the X coordinate, or longitude.

The second bias position is the Y coordinate, or latitude.

For example, bias=xLongitude&bias=yLatitude.

filterCountries :: Maybe (NonEmpty Text)

Limits the search to the given a list of countries/regions. An optional parameter.

Use the ISO 3166 3-digit country code. For example, Australia uses three upper-case characters: AUS.

maxResults :: Maybe Natural

An optional parameter. The maximum number of results returned per request.

The default: 50

indexName :: Text

The name of the place index resource you want to use for the search.

text :: Sensitive Text

The address, name, city, or region to be used in the search. In free-form text format. For example, 123 Any Street.

newSearchPlaceIndexForText

Arguments

:: Text
$sel:indexName:SearchPlaceIndexForText'

-> Text
$sel:text:SearchPlaceIndexForText'

-> SearchPlaceIndexForText

Create a value of SearchPlaceIndexForText with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filterBBox:SearchPlaceIndexForText', searchPlaceIndexForText_filterBBox - Filters the results by returning only Places within the provided bounding box. An optional parameter.

The first 2 bbox parameters describe the lower southwest corner:

The first bbox position is the X coordinate or longitude of the lower southwest corner.

The second bbox position is the Y coordinate or latitude of the lower southwest corner.

For example, bbox=xLongitudeSW&bbox=yLatitudeSW.

The next bbox parameters describe the upper northeast corner:

The third bbox position is the X coordinate, or longitude of the upper northeast corner.

The fourth bbox position is the Y coordinate, or longitude of the upper northeast corner.

For example, bbox=xLongitudeNE&bbox=yLatitudeNE

$sel:biasPosition:SearchPlaceIndexForText', searchPlaceIndexForText_biasPosition - Searches for results closest to the given position. An optional parameter defined by longitude, and latitude.

The first bias position is the X coordinate, or longitude.

The second bias position is the Y coordinate, or latitude.

For example, bias=xLongitude&bias=yLatitude.

$sel:filterCountries:SearchPlaceIndexForText', searchPlaceIndexForText_filterCountries - Limits the search to the given a list of countries/regions. An optional parameter.

Use the ISO 3166 3-digit country code. For example, Australia uses three upper-case characters: AUS.

$sel:maxResults:SearchPlaceIndexForText', searchPlaceIndexForText_maxResults - An optional parameter. The maximum number of results returned per request.

The default: 50

$sel:indexName:SearchPlaceIndexForText', searchPlaceIndexForText_indexName - The name of the place index resource you want to use for the search.

$sel:text:SearchPlaceIndexForText', searchPlaceIndexForText_text - The address, name, city, or region to be used in the search. In free-form text format. For example, 123 Any Street.

Request Lenses

searchPlaceIndexForText_filterBBox :: Lens' SearchPlaceIndexForText (Maybe (NonEmpty Double))

Filters the results by returning only Places within the provided bounding box. An optional parameter.

The first 2 bbox parameters describe the lower southwest corner:

The first bbox position is the X coordinate or longitude of the lower southwest corner.

The second bbox position is the Y coordinate or latitude of the lower southwest corner.

For example, bbox=xLongitudeSW&bbox=yLatitudeSW.

The next bbox parameters describe the upper northeast corner:

The third bbox position is the X coordinate, or longitude of the upper northeast corner.

The fourth bbox position is the Y coordinate, or longitude of the upper northeast corner.

For example, bbox=xLongitudeNE&bbox=yLatitudeNE

searchPlaceIndexForText_biasPosition :: Lens' SearchPlaceIndexForText (Maybe (NonEmpty Double))

Searches for results closest to the given position. An optional parameter defined by longitude, and latitude.

The first bias position is the X coordinate, or longitude.

The second bias position is the Y coordinate, or latitude.

For example, bias=xLongitude&bias=yLatitude.

searchPlaceIndexForText_filterCountries :: Lens' SearchPlaceIndexForText (Maybe (NonEmpty Text))

Limits the search to the given a list of countries/regions. An optional parameter.

Use the ISO 3166 3-digit country code. For example, Australia uses three upper-case characters: AUS.

searchPlaceIndexForText_maxResults :: Lens' SearchPlaceIndexForText (Maybe Natural)

An optional parameter. The maximum number of results returned per request.

The default: 50

searchPlaceIndexForText_indexName :: Lens' SearchPlaceIndexForText Text

The name of the place index resource you want to use for the search.

searchPlaceIndexForText_text :: Lens' SearchPlaceIndexForText Text

The address, name, city, or region to be used in the search. In free-form text format. For example, 123 Any Street.

Destructuring the Response

data SearchPlaceIndexForTextResponse

See: newSearchPlaceIndexForTextResponse smart constructor.

Constructors

SearchPlaceIndexForTextResponse'

Fields

httpStatus :: Int

The response's http status code.

results :: [SearchForTextResult]

A list of Places closest to the specified position. Each result contains additional information about the specific point of interest.

summary :: SearchPlaceIndexForTextSummary

Contains a summary of the request. Contains the BiasPosition, DataSource, FilterBBox, FilterCountries, MaxResults, ResultBBox, and Text.

newSearchPlaceIndexForTextResponse

Arguments

:: Int
$sel:httpStatus:SearchPlaceIndexForTextResponse'

-> SearchPlaceIndexForTextSummary
$sel:summary:SearchPlaceIndexForTextResponse'

-> SearchPlaceIndexForTextResponse

Create a value of SearchPlaceIndexForTextResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:SearchPlaceIndexForTextResponse', searchPlaceIndexForTextResponse_httpStatus - The response's http status code.

$sel:results:SearchPlaceIndexForTextResponse', searchPlaceIndexForTextResponse_results - A list of Places closest to the specified position. Each result contains additional information about the specific point of interest.

$sel:summary:SearchPlaceIndexForTextResponse', searchPlaceIndexForTextResponse_summary - Contains a summary of the request. Contains the BiasPosition, DataSource, FilterBBox, FilterCountries, MaxResults, ResultBBox, and Text.

Response Lenses

searchPlaceIndexForTextResponse_httpStatus :: Lens' SearchPlaceIndexForTextResponse Int

The response's http status code.

searchPlaceIndexForTextResponse_results :: Lens' SearchPlaceIndexForTextResponse [SearchForTextResult]

A list of Places closest to the specified position. Each result contains additional information about the specific point of interest.

searchPlaceIndexForTextResponse_summary :: Lens' SearchPlaceIndexForTextResponse SearchPlaceIndexForTextSummary

Contains a summary of the request. Contains the BiasPosition, DataSource, FilterBBox, FilterCountries, MaxResults, ResultBBox, and Text.