Network.AWS.Location.Types.BatchEvaluateGeofencesError

Description

Documentation

data BatchEvaluateGeofencesError

Contains error details for each device that failed to evaluate its position against the geofences in a given geofence collection.

See: newBatchEvaluateGeofencesError smart constructor.

Constructors

BatchEvaluateGeofencesError'

Fields

deviceId :: Text

The device associated with the position evaluation error.

error :: BatchItemError

Contains details associated to the batch error.

sampleTime :: POSIX

Specifies a timestamp for when the error occurred in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

newBatchEvaluateGeofencesError

Arguments

:: Text
$sel:deviceId:BatchEvaluateGeofencesError'

-> BatchItemError
$sel:error:BatchEvaluateGeofencesError'

-> UTCTime
$sel:sampleTime:BatchEvaluateGeofencesError'

-> BatchEvaluateGeofencesError

Create a value of BatchEvaluateGeofencesError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceId:BatchEvaluateGeofencesError', batchEvaluateGeofencesError_deviceId - The device associated with the position evaluation error.

$sel:error:BatchEvaluateGeofencesError', batchEvaluateGeofencesError_error - Contains details associated to the batch error.

$sel:sampleTime:BatchEvaluateGeofencesError', batchEvaluateGeofencesError_sampleTime - Specifies a timestamp for when the error occurred in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

batchEvaluateGeofencesError_deviceId :: Lens' BatchEvaluateGeofencesError Text

The device associated with the position evaluation error.

batchEvaluateGeofencesError_error :: Lens' BatchEvaluateGeofencesError BatchItemError

Contains details associated to the batch error.

batchEvaluateGeofencesError_sampleTime :: Lens' BatchEvaluateGeofencesError UTCTime

Specifies a timestamp for when the error occurred in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ