Network.AWS.Location.Types.BatchPutGeofenceError

Description

Documentation

data BatchPutGeofenceError

Contains error details for each geofence that failed to be stored in a given geofence collection.

See: newBatchPutGeofenceError smart constructor.

Constructors

BatchPutGeofenceError'

Fields

error :: BatchItemError

Contains details associated to the batch error.

geofenceId :: Text

The geofence associated with the error message.

newBatchPutGeofenceError

Arguments

:: BatchItemError
$sel:error:BatchPutGeofenceError'

-> Text
$sel:geofenceId:BatchPutGeofenceError'

-> BatchPutGeofenceError

Create a value of BatchPutGeofenceError with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:error:BatchPutGeofenceError', batchPutGeofenceError_error - Contains details associated to the batch error.

$sel:geofenceId:BatchPutGeofenceError', batchPutGeofenceError_geofenceId - The geofence associated with the error message.

batchPutGeofenceError_error :: Lens' BatchPutGeofenceError BatchItemError

Contains details associated to the batch error.

batchPutGeofenceError_geofenceId :: Lens' BatchPutGeofenceError Text

The geofence associated with the error message.