Network.AWS.Location.Types.BatchPutGeofenceSuccess

Description

Documentation

data BatchPutGeofenceSuccess

Contains a summary of each geofence that was successfully stored in a given geofence collection.

See: newBatchPutGeofenceSuccess smart constructor.

Constructors

BatchPutGeofenceSuccess'

Fields

createTime :: POSIX

The timestamp for when the geofence was stored in a geofence collection in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

geofenceId :: Text

The geofence successfully stored in a geofence collection.

updateTime :: POSIX

The timestamp for when the geofence was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

newBatchPutGeofenceSuccess

Arguments

:: UTCTime
$sel:createTime:BatchPutGeofenceSuccess'

-> Text
$sel:geofenceId:BatchPutGeofenceSuccess'

-> UTCTime
$sel:updateTime:BatchPutGeofenceSuccess'

-> BatchPutGeofenceSuccess

Create a value of BatchPutGeofenceSuccess with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createTime:BatchPutGeofenceSuccess', batchPutGeofenceSuccess_createTime - The timestamp for when the geofence was stored in a geofence collection in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

$sel:geofenceId:BatchPutGeofenceSuccess', batchPutGeofenceSuccess_geofenceId - The geofence successfully stored in a geofence collection.

$sel:updateTime:BatchPutGeofenceSuccess', batchPutGeofenceSuccess_updateTime - The timestamp for when the geofence was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

batchPutGeofenceSuccess_createTime :: Lens' BatchPutGeofenceSuccess UTCTime

The timestamp for when the geofence was stored in a geofence collection in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

batchPutGeofenceSuccess_geofenceId :: Lens' BatchPutGeofenceSuccess Text

The geofence successfully stored in a geofence collection.

batchPutGeofenceSuccess_updateTime :: Lens' BatchPutGeofenceSuccess UTCTime

The timestamp for when the geofence was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ