Network.AWS.Location.Types.DevicePosition

Description

Documentation

data DevicePosition

Contains the device position details.

See: newDevicePosition smart constructor.

Constructors

DevicePosition'

Fields

deviceId :: Maybe Text

The device whose position you retrieved.

position :: Sensitive (NonEmpty Double)

The last known device position.

receivedTime :: POSIX

The timestamp for when the tracker resource received the device position in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

sampleTime :: POSIX

The timestamp at which the device's position was determined. Uses ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

newDevicePosition

Arguments

:: NonEmpty Double
$sel:position:DevicePosition'

-> UTCTime
$sel:receivedTime:DevicePosition'

-> UTCTime
$sel:sampleTime:DevicePosition'

-> DevicePosition

Create a value of DevicePosition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceId:DevicePosition', devicePosition_deviceId - The device whose position you retrieved.

$sel:position:DevicePosition', devicePosition_position - The last known device position.

$sel:receivedTime:DevicePosition', devicePosition_receivedTime - The timestamp for when the tracker resource received the device position in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

$sel:sampleTime:DevicePosition', devicePosition_sampleTime - The timestamp at which the device's position was determined. Uses ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

devicePosition_deviceId :: Lens' DevicePosition (Maybe Text)

The device whose position you retrieved.

devicePosition_position :: Lens' DevicePosition (NonEmpty Double)

The last known device position.

devicePosition_receivedTime :: Lens' DevicePosition UTCTime

The timestamp for when the tracker resource received the device position in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

devicePosition_sampleTime :: Lens' DevicePosition UTCTime

The timestamp at which the device's position was determined. Uses ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.