Network.AWS.Location.Types.ListDevicePositionsResponseEntry

Description

Documentation

data ListDevicePositionsResponseEntry

Contains the tracker resource details.

See: newListDevicePositionsResponseEntry smart constructor.

Constructors

ListDevicePositionsResponseEntry'

Fields

deviceId :: Text

The ID of the device for this position.

position :: Sensitive (NonEmpty Double)

The last known device position. Empty if no positions currently stored.

sampleTime :: POSIX

The timestamp at which the device position was determined. Uses ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

newListDevicePositionsResponseEntry

Arguments

:: Text
$sel:deviceId:ListDevicePositionsResponseEntry'

-> NonEmpty Double
$sel:position:ListDevicePositionsResponseEntry'

-> UTCTime
$sel:sampleTime:ListDevicePositionsResponseEntry'

-> ListDevicePositionsResponseEntry

Create a value of ListDevicePositionsResponseEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceId:ListDevicePositionsResponseEntry', listDevicePositionsResponseEntry_deviceId - The ID of the device for this position.

$sel:position:ListDevicePositionsResponseEntry', listDevicePositionsResponseEntry_position - The last known device position. Empty if no positions currently stored.

$sel:sampleTime:ListDevicePositionsResponseEntry', listDevicePositionsResponseEntry_sampleTime - The timestamp at which the device position was determined. Uses ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

listDevicePositionsResponseEntry_deviceId :: Lens' ListDevicePositionsResponseEntry Text

The ID of the device for this position.

listDevicePositionsResponseEntry_position :: Lens' ListDevicePositionsResponseEntry (NonEmpty Double)

The last known device position. Empty if no positions currently stored.

listDevicePositionsResponseEntry_sampleTime :: Lens' ListDevicePositionsResponseEntry UTCTime

The timestamp at which the device position was determined. Uses ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.