Network.AWS.Location.Types.ListGeofenceCollectionsResponseEntry

Description

Documentation

data ListGeofenceCollectionsResponseEntry

Contains the geofence collection details.

See: newListGeofenceCollectionsResponseEntry smart constructor.

Constructors

ListGeofenceCollectionsResponseEntry'

Fields

pricingPlanDataSource :: Maybe Text

The specified data provider for the geofence collection.

collectionName :: Text

The name of the geofence collection.

createTime :: POSIX

The timestamp for when the geofence collection was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

description :: Text

The description for the geofence collection

pricingPlan :: PricingPlan

The pricing plan for the specified geofence collection.

For additional details and restrictions on each pricing plan option, see the Amazon Location Service pricing page.

updateTime :: POSIX

Specifies a timestamp for when the resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

newListGeofenceCollectionsResponseEntry

Arguments

:: Text
$sel:collectionName:ListGeofenceCollectionsResponseEntry'

-> UTCTime
$sel:createTime:ListGeofenceCollectionsResponseEntry'

-> Text
$sel:description:ListGeofenceCollectionsResponseEntry'

-> PricingPlan
$sel:pricingPlan:ListGeofenceCollectionsResponseEntry'

-> UTCTime
$sel:updateTime:ListGeofenceCollectionsResponseEntry'

-> ListGeofenceCollectionsResponseEntry

Create a value of ListGeofenceCollectionsResponseEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pricingPlanDataSource:ListGeofenceCollectionsResponseEntry', listGeofenceCollectionsResponseEntry_pricingPlanDataSource - The specified data provider for the geofence collection.

$sel:collectionName:ListGeofenceCollectionsResponseEntry', listGeofenceCollectionsResponseEntry_collectionName - The name of the geofence collection.

$sel:createTime:ListGeofenceCollectionsResponseEntry', listGeofenceCollectionsResponseEntry_createTime - The timestamp for when the geofence collection was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

$sel:description:ListGeofenceCollectionsResponseEntry', listGeofenceCollectionsResponseEntry_description - The description for the geofence collection

$sel:pricingPlan:ListGeofenceCollectionsResponseEntry', listGeofenceCollectionsResponseEntry_pricingPlan - The pricing plan for the specified geofence collection.

For additional details and restrictions on each pricing plan option, see the Amazon Location Service pricing page.

$sel:updateTime:ListGeofenceCollectionsResponseEntry', listGeofenceCollectionsResponseEntry_updateTime - Specifies a timestamp for when the resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

listGeofenceCollectionsResponseEntry_pricingPlanDataSource :: Lens' ListGeofenceCollectionsResponseEntry (Maybe Text)

The specified data provider for the geofence collection.

listGeofenceCollectionsResponseEntry_collectionName :: Lens' ListGeofenceCollectionsResponseEntry Text

The name of the geofence collection.

listGeofenceCollectionsResponseEntry_createTime :: Lens' ListGeofenceCollectionsResponseEntry UTCTime

The timestamp for when the geofence collection was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

listGeofenceCollectionsResponseEntry_description :: Lens' ListGeofenceCollectionsResponseEntry Text

The description for the geofence collection

listGeofenceCollectionsResponseEntry_pricingPlan :: Lens' ListGeofenceCollectionsResponseEntry PricingPlan

The pricing plan for the specified geofence collection.

For additional details and restrictions on each pricing plan option, see the Amazon Location Service pricing page.

listGeofenceCollectionsResponseEntry_updateTime :: Lens' ListGeofenceCollectionsResponseEntry UTCTime

Specifies a timestamp for when the resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ