Network.AWS.Location.Types.ListGeofenceResponseEntry

Description

Documentation

data ListGeofenceResponseEntry

Contains a list of geofences stored in a given geofence collection.

See: newListGeofenceResponseEntry smart constructor.

Constructors

ListGeofenceResponseEntry'

Fields

createTime :: POSIX

The timestamp for when the geofence was stored in a geofence collection in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

geofenceId :: Text

The geofence identifier.

geometry :: GeofenceGeometry

Contains the geofence geometry details describing a polygon.

status :: Text

Identifies the state of the geofence. A geofence will hold one of the following states:

ACTIVE — The geofence has been indexed by the system.

PENDING — The geofence is being processed by the system.

FAILED — The geofence failed to be indexed by the system.

DELETED — The geofence has been deleted from the system index.

DELETING — The geofence is being deleted from the system index.

updateTime :: POSIX

The timestamp for when the geofence was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

newListGeofenceResponseEntry

Arguments

:: UTCTime
$sel:createTime:ListGeofenceResponseEntry'

-> Text
$sel:geofenceId:ListGeofenceResponseEntry'

-> GeofenceGeometry
$sel:geometry:ListGeofenceResponseEntry'

-> Text
$sel:status:ListGeofenceResponseEntry'

-> UTCTime
$sel:updateTime:ListGeofenceResponseEntry'

-> ListGeofenceResponseEntry

Create a value of ListGeofenceResponseEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createTime:ListGeofenceResponseEntry', listGeofenceResponseEntry_createTime - The timestamp for when the geofence was stored in a geofence collection in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

$sel:geofenceId:ListGeofenceResponseEntry', listGeofenceResponseEntry_geofenceId - The geofence identifier.

$sel:geometry:ListGeofenceResponseEntry', listGeofenceResponseEntry_geometry - Contains the geofence geometry details describing a polygon.

$sel:status:ListGeofenceResponseEntry', listGeofenceResponseEntry_status - Identifies the state of the geofence. A geofence will hold one of the following states:

ACTIVE — The geofence has been indexed by the system.

PENDING — The geofence is being processed by the system.

FAILED — The geofence failed to be indexed by the system.

DELETED — The geofence has been deleted from the system index.

DELETING — The geofence is being deleted from the system index.

$sel:updateTime:ListGeofenceResponseEntry', listGeofenceResponseEntry_updateTime - The timestamp for when the geofence was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

listGeofenceResponseEntry_createTime :: Lens' ListGeofenceResponseEntry UTCTime

The timestamp for when the geofence was stored in a geofence collection in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ

listGeofenceResponseEntry_geofenceId :: Lens' ListGeofenceResponseEntry Text

The geofence identifier.

listGeofenceResponseEntry_geometry :: Lens' ListGeofenceResponseEntry GeofenceGeometry

Contains the geofence geometry details describing a polygon.

listGeofenceResponseEntry_status :: Lens' ListGeofenceResponseEntry Text

Identifies the state of the geofence. A geofence will hold one of the following states:

ACTIVE — The geofence has been indexed by the system.

PENDING — The geofence is being processed by the system.

FAILED — The geofence failed to be indexed by the system.

DELETED — The geofence has been deleted from the system index.

DELETING — The geofence is being deleted from the system index.

listGeofenceResponseEntry_updateTime :: Lens' ListGeofenceResponseEntry UTCTime

The timestamp for when the geofence was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ