Network.AWS.Location.Types.ListPlaceIndexesResponseEntry

Description

Documentation

data ListPlaceIndexesResponseEntry

A place index resource listed in your AWS account.

See: newListPlaceIndexesResponseEntry smart constructor.

Constructors

ListPlaceIndexesResponseEntry'

Fields

createTime :: POSIX

The timestamp for when the place index resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

dataSource :: Text

The data provider of geospatial data. Indicates one of the available providers:

Esri

Here

For additional details on data providers, see Amazon Location Service data providers.

description :: Text

The optional description for the place index resource.

indexName :: Text

The name of the place index resource.

pricingPlan :: PricingPlan

The pricing plan for the specified place index resource.

For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

updateTime :: POSIX

The timestamp for when the place index resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

newListPlaceIndexesResponseEntry

Arguments

:: UTCTime
$sel:createTime:ListPlaceIndexesResponseEntry'

-> Text
$sel:dataSource:ListPlaceIndexesResponseEntry'

-> Text
$sel:description:ListPlaceIndexesResponseEntry'

-> Text
$sel:indexName:ListPlaceIndexesResponseEntry'

-> PricingPlan
$sel:pricingPlan:ListPlaceIndexesResponseEntry'

-> UTCTime
$sel:updateTime:ListPlaceIndexesResponseEntry'

-> ListPlaceIndexesResponseEntry

Create a value of ListPlaceIndexesResponseEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createTime:ListPlaceIndexesResponseEntry', listPlaceIndexesResponseEntry_createTime - The timestamp for when the place index resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

$sel:dataSource:ListPlaceIndexesResponseEntry', listPlaceIndexesResponseEntry_dataSource - The data provider of geospatial data. Indicates one of the available providers:

Esri

Here

For additional details on data providers, see Amazon Location Service data providers.

$sel:description:ListPlaceIndexesResponseEntry', listPlaceIndexesResponseEntry_description - The optional description for the place index resource.

$sel:indexName:ListPlaceIndexesResponseEntry', listPlaceIndexesResponseEntry_indexName - The name of the place index resource.

$sel:pricingPlan:ListPlaceIndexesResponseEntry', listPlaceIndexesResponseEntry_pricingPlan - The pricing plan for the specified place index resource.

For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

$sel:updateTime:ListPlaceIndexesResponseEntry', listPlaceIndexesResponseEntry_updateTime - The timestamp for when the place index resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

listPlaceIndexesResponseEntry_createTime :: Lens' ListPlaceIndexesResponseEntry UTCTime

The timestamp for when the place index resource was created in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

listPlaceIndexesResponseEntry_dataSource :: Lens' ListPlaceIndexesResponseEntry Text

The data provider of geospatial data. Indicates one of the available providers:

Esri

Here

For additional details on data providers, see Amazon Location Service data providers.

listPlaceIndexesResponseEntry_description :: Lens' ListPlaceIndexesResponseEntry Text

The optional description for the place index resource.

listPlaceIndexesResponseEntry_indexName :: Lens' ListPlaceIndexesResponseEntry Text

The name of the place index resource.

listPlaceIndexesResponseEntry_pricingPlan :: Lens' ListPlaceIndexesResponseEntry PricingPlan

The pricing plan for the specified place index resource.

For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

listPlaceIndexesResponseEntry_updateTime :: Lens' ListPlaceIndexesResponseEntry UTCTime

The timestamp for when the place index resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.