Network.AWS.Location.Types.Place

Description

Documentation

data Place

Contains details about addresses or points of interest that match the search criteria.

See: newPlace smart constructor.

Constructors

Place'

Fields

municipality :: Maybe Text

A name for a local area, such as a city or town name. For example, Toronto.

addressNumber :: Maybe Text

The numerical portion of an address, such as a building number.

postalCode :: Maybe Text

A group of numbers and letters in a country-specific format, which accompanies the address for the purpose of identifying a location.

country :: Maybe Text

A country/region specified using ISO 3166 3-digit country/region code. For example, CAN.

street :: Maybe Text

The name for a street or a road to identify a location. For example, Main Street.

subRegion :: Maybe Text

A country, or an area that's part of a larger region . For example, Metro Vancouver.

region :: Maybe Text

A name for an area or geographical division, such as a province or state name. For example, British Columbia.

label :: Maybe Text

The full name and address of the point of interest such as a city, region, or country. For example, 123 Any Street, Any Town, USA.

neighborhood :: Maybe Text

The name of a community district. For example, Downtown.

geometry :: PlaceGeometry

newPlace

Arguments

:: PlaceGeometry
$sel:geometry:Place'

-> Place

Create a value of Place with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:municipality:Place', place_municipality - A name for a local area, such as a city or town name. For example, Toronto.

$sel:addressNumber:Place', place_addressNumber - The numerical portion of an address, such as a building number.

$sel:postalCode:Place', place_postalCode - A group of numbers and letters in a country-specific format, which accompanies the address for the purpose of identifying a location.

$sel:country:Place', place_country - A country/region specified using ISO 3166 3-digit country/region code. For example, CAN.

$sel:street:Place', place_street - The name for a street or a road to identify a location. For example, Main Street.

$sel:subRegion:Place', place_subRegion - A country, or an area that's part of a larger region . For example, Metro Vancouver.

$sel:region:Place', place_region - A name for an area or geographical division, such as a province or state name. For example, British Columbia.

$sel:label:Place', place_label - The full name and address of the point of interest such as a city, region, or country. For example, 123 Any Street, Any Town, USA.

$sel:neighborhood:Place', place_neighborhood - The name of a community district. For example, Downtown.

$sel:geometry:Place', place_geometry - Undocumented member.

place_municipality :: Lens' Place (Maybe Text)

A name for a local area, such as a city or town name. For example, Toronto.

place_addressNumber :: Lens' Place (Maybe Text)

The numerical portion of an address, such as a building number.

place_postalCode :: Lens' Place (Maybe Text)

A group of numbers and letters in a country-specific format, which accompanies the address for the purpose of identifying a location.

place_country :: Lens' Place (Maybe Text)

A country/region specified using ISO 3166 3-digit country/region code. For example, CAN.

place_street :: Lens' Place (Maybe Text)

The name for a street or a road to identify a location. For example, Main Street.

place_subRegion :: Lens' Place (Maybe Text)

A country, or an area that's part of a larger region . For example, Metro Vancouver.

place_region :: Lens' Place (Maybe Text)

A name for an area or geographical division, such as a province or state name. For example, British Columbia.

place_label :: Lens' Place (Maybe Text)

The full name and address of the point of interest such as a city, region, or country. For example, 123 Any Street, Any Town, USA.

place_neighborhood :: Lens' Place (Maybe Text)

The name of a community district. For example, Downtown.

place_geometry :: Lens' Place PlaceGeometry

Undocumented member.