Network.AWS.Location.Types.Step

Description

Documentation

data Step

Represents an element of a leg within a route. A step contains instructions for how to move to the next step in the leg.

See: newStep smart constructor.

Constructors

Step'

Fields

geometryOffset :: Maybe Natural

Represents the start position, or index, in a sequence of steps within the leg's line string geometry. For example, the index of the first step in a leg geometry is 0.

Included in the response for queries that set IncludeLegGeometry to True.

distance :: Double

The travel distance between the step's StartPosition and EndPosition.

durationSeconds :: Double

The estimated travel time, in seconds, from the step's StartPosition to the EndPosition. . The travel mode and departure time that you specify in the request determines the calculated time.

endPosition :: Sensitive (NonEmpty Double)

The end position of a step. If the position the last step in the leg, this position is the same as the end position of the leg.

startPosition :: Sensitive (NonEmpty Double)

The starting position of a step. If the position is the first step in the leg, this position is the same as the start position of the leg.

newStep

Arguments

:: Double
$sel:distance:Step'

-> Double
$sel:durationSeconds:Step'

-> NonEmpty Double
$sel:endPosition:Step'

-> NonEmpty Double
$sel:startPosition:Step'

-> Step

Create a value of Step with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:geometryOffset:Step', step_geometryOffset - Represents the start position, or index, in a sequence of steps within the leg's line string geometry. For example, the index of the first step in a leg geometry is 0.

Included in the response for queries that set IncludeLegGeometry to True.

$sel:distance:Step', step_distance - The travel distance between the step's StartPosition and EndPosition.

$sel:durationSeconds:Step', step_durationSeconds - The estimated travel time, in seconds, from the step's StartPosition to the EndPosition. . The travel mode and departure time that you specify in the request determines the calculated time.

$sel:endPosition:Step', step_endPosition - The end position of a step. If the position the last step in the leg, this position is the same as the end position of the leg.

$sel:startPosition:Step', step_startPosition - The starting position of a step. If the position is the first step in the leg, this position is the same as the start position of the leg.

step_geometryOffset :: Lens' Step (Maybe Natural)

Represents the start position, or index, in a sequence of steps within the leg's line string geometry. For example, the index of the first step in a leg geometry is 0.

Included in the response for queries that set IncludeLegGeometry to True.

step_distance :: Lens' Step Double

The travel distance between the step's StartPosition and EndPosition.

step_durationSeconds :: Lens' Step Double

The estimated travel time, in seconds, from the step's StartPosition to the EndPosition. . The travel mode and departure time that you specify in the request determines the calculated time.

step_endPosition :: Lens' Step (NonEmpty Double)

The end position of a step. If the position the last step in the leg, this position is the same as the end position of the leg.

step_startPosition :: Lens' Step (NonEmpty Double)

The starting position of a step. If the position is the first step in the leg, this position is the same as the start position of the leg.