Network.AWS.Location.Types.TruckWeight

Description

Documentation

data TruckWeight

Contains details about the truck's weight specifications. Used to avoid roads that can't support or allow the total weight for requests that specify TravelMode as Truck.

See: newTruckWeight smart constructor.

Constructors

TruckWeight'

Fields

total :: Maybe Double

The total weight of the truck.

For example, 3500.

unit :: Maybe VehicleWeightUnit

The unit of measurement to use for the truck weight.

Default Value: Kilograms

newTruckWeight :: TruckWeight

Create a value of TruckWeight with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:total:TruckWeight', truckWeight_total - The total weight of the truck.

For example, 3500.

$sel:unit:TruckWeight', truckWeight_unit - The unit of measurement to use for the truck weight.

Default Value: Kilograms

truckWeight_total :: Lens' TruckWeight (Maybe Double)

The total weight of the truck.

For example, 3500.

truckWeight_unit :: Lens' TruckWeight (Maybe VehicleWeightUnit)

The unit of measurement to use for the truck weight.

Default Value: Kilograms