Network.AWS.Location.UpdatePlaceIndex

Description

Updates the specified properties of a given place index resource.

Creating a Request

data UpdatePlaceIndex

See: newUpdatePlaceIndex smart constructor.

Constructors

UpdatePlaceIndex'

Fields

pricingPlan :: Maybe PricingPlan

Updates the pricing plan for the place index resource.

For more information about each pricing plan option restrictions, see Amazon Location Service pricing.

dataSourceConfiguration :: Maybe DataSourceConfiguration

Updates the data storage option for the place index resource.

description :: Maybe Text

Updates the description for the place index resource.

indexName :: Text

The name of the place index resource to update.

newUpdatePlaceIndex

Arguments

:: Text
$sel:indexName:UpdatePlaceIndex'

-> UpdatePlaceIndex

Create a value of UpdatePlaceIndex with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pricingPlan:UpdatePlaceIndex', updatePlaceIndex_pricingPlan - Updates the pricing plan for the place index resource.

For more information about each pricing plan option restrictions, see Amazon Location Service pricing.

$sel:dataSourceConfiguration:UpdatePlaceIndex', updatePlaceIndex_dataSourceConfiguration - Updates the data storage option for the place index resource.

$sel:description:UpdatePlaceIndex', updatePlaceIndex_description - Updates the description for the place index resource.

$sel:indexName:UpdatePlaceIndex', updatePlaceIndex_indexName - The name of the place index resource to update.

Request Lenses

updatePlaceIndex_pricingPlan :: Lens' UpdatePlaceIndex (Maybe PricingPlan)

Updates the pricing plan for the place index resource.

For more information about each pricing plan option restrictions, see Amazon Location Service pricing.

updatePlaceIndex_dataSourceConfiguration :: Lens' UpdatePlaceIndex (Maybe DataSourceConfiguration)

Updates the data storage option for the place index resource.

updatePlaceIndex_description :: Lens' UpdatePlaceIndex (Maybe Text)

Updates the description for the place index resource.

updatePlaceIndex_indexName :: Lens' UpdatePlaceIndex Text

The name of the place index resource to update.

Destructuring the Response

data UpdatePlaceIndexResponse

See: newUpdatePlaceIndexResponse smart constructor.

Constructors

UpdatePlaceIndexResponse'

Fields

httpStatus :: Int

The response's http status code.

indexArn :: Text

The Amazon Resource Name (ARN) of the upated place index resource. Used to specify a resource across AWS.

Format example: arn:aws:geo:region:account-id:place- index/ExamplePlaceIndex

indexName :: Text

The name of the updated place index resource.

updateTime :: POSIX

The timestamp for when the place index resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

newUpdatePlaceIndexResponse

Arguments

:: Int
$sel:httpStatus:UpdatePlaceIndexResponse'

-> Text
$sel:indexArn:UpdatePlaceIndexResponse'

-> Text
$sel:indexName:UpdatePlaceIndex'

-> UTCTime
$sel:updateTime:UpdatePlaceIndexResponse'

-> UpdatePlaceIndexResponse

Create a value of UpdatePlaceIndexResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdatePlaceIndexResponse', updatePlaceIndexResponse_httpStatus - The response's http status code.

$sel:indexArn:UpdatePlaceIndexResponse', updatePlaceIndexResponse_indexArn - The Amazon Resource Name (ARN) of the upated place index resource. Used to specify a resource across AWS.

Format example: arn:aws:geo:region:account-id:place- index/ExamplePlaceIndex

$sel:indexName:UpdatePlaceIndex', updatePlaceIndexResponse_indexName - The name of the updated place index resource.

$sel:updateTime:UpdatePlaceIndexResponse', updatePlaceIndexResponse_updateTime - The timestamp for when the place index resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.

Response Lenses

updatePlaceIndexResponse_httpStatus :: Lens' UpdatePlaceIndexResponse Int

The response's http status code.

updatePlaceIndexResponse_indexArn :: Lens' UpdatePlaceIndexResponse Text

The Amazon Resource Name (ARN) of the upated place index resource. Used to specify a resource across AWS.

Format example: arn:aws:geo:region:account-id:place- index/ExamplePlaceIndex

updatePlaceIndexResponse_indexName :: Lens' UpdatePlaceIndexResponse Text

The name of the updated place index resource.

updatePlaceIndexResponse_updateTime :: Lens' UpdatePlaceIndexResponse UTCTime

The timestamp for when the place index resource was last updated in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ.