Network.AWS.Logger

Description

Types and functions for constructing loggers and emitting log messages.

Constructing a Logger

type Logger = LogLevel -> ByteStringBuilder -> IO ()

A function threaded through various request and serialisation routines to log informational and debug messages.

newLogger :: MonadIO m => LogLevel -> Handle -> m Logger

This is a primitive logger which can be used to log builds to a Handle.

Note: A more sophisticated logging library such as tinylog or fast-logger should be used in production code.

Levels

data LogLevel

Constructors

Info
Info messages supplied by the user - this level is not emitted by the library.

Error
Error messages only.

Debug
Useful debug information + info + error levels.

Trace
Includes potentially sensitive signing metadata, and non-streaming response bodies.

logError :: (MonadIO m, ToLog a) => Logger -> a -> m ()

logInfo :: (MonadIO m, ToLog a) => Logger -> a -> m ()

logDebug :: (MonadIO m, ToLog a) => Logger -> a -> m ()

logTrace :: (MonadIO m, ToLog a) => Logger -> a -> m ()

Building Messages

class ToLog a where

Methods

build :: a -> ByteStringBuilder

Convert a value to a loggable builder.

buildLines :: [ByteStringBuilder] -> ByteStringBuilder

Intercalate a list of ByteStringBuilders with newlines.