Network.AWS.LookoutEquipment.CreateInferenceScheduler

Description

Creates a scheduled inference. Scheduling an inference is setting up a continuous real-time inference plan to analyze new measurement data. When setting up the schedule, you provide an S3 bucket location for the input data, assign it a delimiter between separate entries in the data, set an offset delay if desired, and set the frequency of inferencing. You must also provide an S3 bucket location for the output data.

Creating a Request

data CreateInferenceScheduler

See: newCreateInferenceScheduler smart constructor.

Constructors

CreateInferenceScheduler'

Fields

dataDelayOffsetInMinutes :: Maybe Natural

A period of time (in minutes) by which inference on the data is delayed after the data starts. For instance, if you select an offset delay time of five minutes, inference will not begin on the data until the first data measurement after the five minute mark. For example, if five minutes is selected, the inference scheduler will wake up at the configured frequency with the additional five minute delay time to check the customer S3 bucket. The customer can upload data at the same frequency and they don't need to stop and restart the scheduler when uploading new data.

tags :: Maybe [Tag]

Any tags associated with the inference scheduler.

serverSideKmsKeyId :: Maybe Text

Provides the identifier of the KMS key used to encrypt inference scheduler data by Amazon Lookout for Equipment.

modelName :: Text

The name of the previously trained ML model being used to create the inference scheduler.

inferenceSchedulerName :: Text

The name of the inference scheduler being created.

dataUploadFrequency :: DataUploadFrequency

How often data is uploaded to the source S3 bucket for the input data. The value chosen is the length of time between data uploads. For instance, if you select 5 minutes, Amazon Lookout for Equipment will upload the real-time data to the source bucket once every 5 minutes. This frequency also determines how often Amazon Lookout for Equipment starts a scheduled inference on your data. In this example, it starts once every 5 minutes.

dataInputConfiguration :: InferenceInputConfiguration

Specifies configuration information for the input data for the inference scheduler, including delimiter, format, and dataset location.

dataOutputConfiguration :: InferenceOutputConfiguration

Specifies configuration information for the output results for the inference scheduler, including the S3 location for the output.

roleArn :: Text

The Amazon Resource Name (ARN) of a role with permission to access the data source being used for the inference.

clientToken :: Text

A unique identifier for the request. If you do not set the client request token, Amazon Lookout for Equipment generates one.

newCreateInferenceScheduler

Arguments

:: Text
$sel:modelName:CreateInferenceScheduler'

-> Text
$sel:inferenceSchedulerName:CreateInferenceScheduler'

-> DataUploadFrequency
$sel:dataUploadFrequency:CreateInferenceScheduler'

-> InferenceInputConfiguration
$sel:dataInputConfiguration:CreateInferenceScheduler'

-> InferenceOutputConfiguration
$sel:dataOutputConfiguration:CreateInferenceScheduler'

-> Text
$sel:roleArn:CreateInferenceScheduler'

-> Text
$sel:clientToken:CreateInferenceScheduler'

-> CreateInferenceScheduler

Create a value of CreateInferenceScheduler with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataDelayOffsetInMinutes:CreateInferenceScheduler', createInferenceScheduler_dataDelayOffsetInMinutes - A period of time (in minutes) by which inference on the data is delayed after the data starts. For instance, if you select an offset delay time of five minutes, inference will not begin on the data until the first data measurement after the five minute mark. For example, if five minutes is selected, the inference scheduler will wake up at the configured frequency with the additional five minute delay time to check the customer S3 bucket. The customer can upload data at the same frequency and they don't need to stop and restart the scheduler when uploading new data.

$sel:tags:CreateInferenceScheduler', createInferenceScheduler_tags - Any tags associated with the inference scheduler.

$sel:serverSideKmsKeyId:CreateInferenceScheduler', createInferenceScheduler_serverSideKmsKeyId - Provides the identifier of the KMS key used to encrypt inference scheduler data by Amazon Lookout for Equipment.

$sel:modelName:CreateInferenceScheduler', createInferenceScheduler_modelName - The name of the previously trained ML model being used to create the inference scheduler.

$sel:inferenceSchedulerName:CreateInferenceScheduler', createInferenceScheduler_inferenceSchedulerName - The name of the inference scheduler being created.

$sel:dataUploadFrequency:CreateInferenceScheduler', createInferenceScheduler_dataUploadFrequency - How often data is uploaded to the source S3 bucket for the input data. The value chosen is the length of time between data uploads. For instance, if you select 5 minutes, Amazon Lookout for Equipment will upload the real-time data to the source bucket once every 5 minutes. This frequency also determines how often Amazon Lookout for Equipment starts a scheduled inference on your data. In this example, it starts once every 5 minutes.

$sel:dataInputConfiguration:CreateInferenceScheduler', createInferenceScheduler_dataInputConfiguration - Specifies configuration information for the input data for the inference scheduler, including delimiter, format, and dataset location.

$sel:dataOutputConfiguration:CreateInferenceScheduler', createInferenceScheduler_dataOutputConfiguration - Specifies configuration information for the output results for the inference scheduler, including the S3 location for the output.

$sel:roleArn:CreateInferenceScheduler', createInferenceScheduler_roleArn - The Amazon Resource Name (ARN) of a role with permission to access the data source being used for the inference.

$sel:clientToken:CreateInferenceScheduler', createInferenceScheduler_clientToken - A unique identifier for the request. If you do not set the client request token, Amazon Lookout for Equipment generates one.

Request Lenses

createInferenceScheduler_dataDelayOffsetInMinutes :: Lens' CreateInferenceScheduler (Maybe Natural)

A period of time (in minutes) by which inference on the data is delayed after the data starts. For instance, if you select an offset delay time of five minutes, inference will not begin on the data until the first data measurement after the five minute mark. For example, if five minutes is selected, the inference scheduler will wake up at the configured frequency with the additional five minute delay time to check the customer S3 bucket. The customer can upload data at the same frequency and they don't need to stop and restart the scheduler when uploading new data.

createInferenceScheduler_tags :: Lens' CreateInferenceScheduler (Maybe [Tag])

Any tags associated with the inference scheduler.

createInferenceScheduler_serverSideKmsKeyId :: Lens' CreateInferenceScheduler (Maybe Text)

Provides the identifier of the KMS key used to encrypt inference scheduler data by Amazon Lookout for Equipment.

createInferenceScheduler_modelName :: Lens' CreateInferenceScheduler Text

The name of the previously trained ML model being used to create the inference scheduler.

createInferenceScheduler_inferenceSchedulerName :: Lens' CreateInferenceScheduler Text

The name of the inference scheduler being created.

createInferenceScheduler_dataUploadFrequency :: Lens' CreateInferenceScheduler DataUploadFrequency

How often data is uploaded to the source S3 bucket for the input data. The value chosen is the length of time between data uploads. For instance, if you select 5 minutes, Amazon Lookout for Equipment will upload the real-time data to the source bucket once every 5 minutes. This frequency also determines how often Amazon Lookout for Equipment starts a scheduled inference on your data. In this example, it starts once every 5 minutes.

createInferenceScheduler_dataInputConfiguration :: Lens' CreateInferenceScheduler InferenceInputConfiguration

Specifies configuration information for the input data for the inference scheduler, including delimiter, format, and dataset location.

createInferenceScheduler_dataOutputConfiguration :: Lens' CreateInferenceScheduler InferenceOutputConfiguration

Specifies configuration information for the output results for the inference scheduler, including the S3 location for the output.

createInferenceScheduler_roleArn :: Lens' CreateInferenceScheduler Text

The Amazon Resource Name (ARN) of a role with permission to access the data source being used for the inference.

createInferenceScheduler_clientToken :: Lens' CreateInferenceScheduler Text

A unique identifier for the request. If you do not set the client request token, Amazon Lookout for Equipment generates one.

Destructuring the Response

data CreateInferenceSchedulerResponse

See: newCreateInferenceSchedulerResponse smart constructor.

Constructors

CreateInferenceSchedulerResponse'

Fields

status :: Maybe InferenceSchedulerStatus

Indicates the status of the CreateInferenceScheduler operation.

inferenceSchedulerArn :: Maybe Text

The Amazon Resource Name (ARN) of the inference scheduler being created.

inferenceSchedulerName :: Maybe Text

The name of inference scheduler being created.

httpStatus :: Int

The response's http status code.

newCreateInferenceSchedulerResponse

Arguments

:: Int
$sel:httpStatus:CreateInferenceSchedulerResponse'

-> CreateInferenceSchedulerResponse

Create a value of CreateInferenceSchedulerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:CreateInferenceSchedulerResponse', createInferenceSchedulerResponse_status - Indicates the status of the CreateInferenceScheduler operation.

$sel:inferenceSchedulerArn:CreateInferenceSchedulerResponse', createInferenceSchedulerResponse_inferenceSchedulerArn - The Amazon Resource Name (ARN) of the inference scheduler being created.

$sel:inferenceSchedulerName:CreateInferenceScheduler', createInferenceSchedulerResponse_inferenceSchedulerName - The name of inference scheduler being created.

$sel:httpStatus:CreateInferenceSchedulerResponse', createInferenceSchedulerResponse_httpStatus - The response's http status code.

Response Lenses

createInferenceSchedulerResponse_status :: Lens' CreateInferenceSchedulerResponse (Maybe InferenceSchedulerStatus)

Indicates the status of the CreateInferenceScheduler operation.

createInferenceSchedulerResponse_inferenceSchedulerArn :: Lens' CreateInferenceSchedulerResponse (Maybe Text)

The Amazon Resource Name (ARN) of the inference scheduler being created.

createInferenceSchedulerResponse_inferenceSchedulerName :: Lens' CreateInferenceSchedulerResponse (Maybe Text)

The name of inference scheduler being created.

createInferenceSchedulerResponse_httpStatus :: Lens' CreateInferenceSchedulerResponse Int

The response's http status code.