Network.AWS.LookoutEquipment.DescribeModel

Description

Provides a JSON containing the overall information about a specific ML model, including model name and ARN, dataset, training and evaluation information, status, and so on.

Creating a Request

data DescribeModel

See: newDescribeModel smart constructor.

Constructors

DescribeModel'

Fields

modelName :: Text

The name of the ML model to be described.

newDescribeModel

Arguments

:: Text
$sel:modelName:DescribeModel'

-> DescribeModel

Create a value of DescribeModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:modelName:DescribeModel', describeModel_modelName - The name of the ML model to be described.

Request Lenses

describeModel_modelName :: Lens' DescribeModel Text

The name of the ML model to be described.

Destructuring the Response

data DescribeModelResponse

See: newDescribeModelResponse smart constructor.

Constructors

DescribeModelResponse'

Fields

status :: Maybe ModelStatus

Specifies the current status of the model being described. Status describes the status of the most recent action of the model.

dataPreProcessingConfiguration :: Maybe DataPreProcessingConfiguration

The configuration is the TargetSamplingRate, which is the sampling rate of the data after post processing by Amazon Lookout for Equipment. For example, if you provide data that has been collected at a 1 second level and you want the system to resample the data at a 1 minute rate before training, the TargetSamplingRate is 1 minute.

When providing a value for the TargetSamplingRate, you must attach the prefix "PT" to the rate you want. The value for a 1 second rate is therefore PT1S, the value for a 15 minute rate is PT15M, and the value for a 1 hour rate is PT1H

trainingExecutionStartTime :: Maybe POSIX

Indicates the time at which the training of the ML model began.

datasetArn :: Maybe Text

The Amazon Resouce Name (ARN) of the dataset used to create the ML model being described.

failedReason :: Maybe Text

If the training of the ML model failed, this indicates the reason for that failure.

modelArn :: Maybe Text

The Amazon Resource Name (ARN) of the ML model being described.

lastUpdatedTime :: Maybe POSIX

Indicates the last time the ML model was updated. The type of update is not specified.

trainingDataEndTime :: Maybe POSIX

Indicates the time reference in the dataset that was used to end the subset of training data for the ML model.

createdAt :: Maybe POSIX

Indicates the time and date at which the ML model was created.

modelName :: Maybe Text

The name of the ML model being described.

modelMetrics :: Maybe Text

The Model Metrics show an aggregated summary of the model's performance within the evaluation time range. This is the JSON content of the metrics created when evaluating the model.

evaluationDataStartTime :: Maybe POSIX

Indicates the time reference in the dataset that was used to begin the subset of evaluation data for the ML model.

schema :: Maybe Text

A JSON description of the data that is in each time series dataset, including names, column names, and data types.

offCondition :: Maybe Text

Indicates that the asset associated with this sensor has been shut off. As long as this condition is met, Lookout for Equipment will not use data from this asset for training, evaluation, or inference.

evaluationDataEndTime :: Maybe POSIX

Indicates the time reference in the dataset that was used to end the subset of evaluation data for the ML model.

datasetName :: Maybe Text

The name of the dataset being used by the ML being described.

trainingDataStartTime :: Maybe POSIX

Indicates the time reference in the dataset that was used to begin the subset of training data for the ML model.

trainingExecutionEndTime :: Maybe POSIX

Indicates the time at which the training of the ML model was completed.

labelsInputConfiguration :: Maybe LabelsInputConfiguration

Specifies configuration information about the labels input, including its S3 location.

serverSideKmsKeyId :: Maybe Text

Provides the identifier of the KMS key used to encrypt model data by Amazon Lookout for Equipment.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of a role with permission to access the data source for the ML model being described.

httpStatus :: Int

The response's http status code.

newDescribeModelResponse

Arguments

:: Int
$sel:httpStatus:DescribeModelResponse'

-> DescribeModelResponse

Create a value of DescribeModelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeModelResponse', describeModelResponse_status - Specifies the current status of the model being described. Status describes the status of the most recent action of the model.

$sel:dataPreProcessingConfiguration:DescribeModelResponse', describeModelResponse_dataPreProcessingConfiguration - The configuration is the TargetSamplingRate, which is the sampling rate of the data after post processing by Amazon Lookout for Equipment. For example, if you provide data that has been collected at a 1 second level and you want the system to resample the data at a 1 minute rate before training, the TargetSamplingRate is 1 minute.

When providing a value for the TargetSamplingRate, you must attach the prefix "PT" to the rate you want. The value for a 1 second rate is therefore PT1S, the value for a 15 minute rate is PT15M, and the value for a 1 hour rate is PT1H

$sel:trainingExecutionStartTime:DescribeModelResponse', describeModelResponse_trainingExecutionStartTime - Indicates the time at which the training of the ML model began.

$sel:datasetArn:DescribeModelResponse', describeModelResponse_datasetArn - The Amazon Resouce Name (ARN) of the dataset used to create the ML model being described.

$sel:failedReason:DescribeModelResponse', describeModelResponse_failedReason - If the training of the ML model failed, this indicates the reason for that failure.

$sel:modelArn:DescribeModelResponse', describeModelResponse_modelArn - The Amazon Resource Name (ARN) of the ML model being described.

$sel:lastUpdatedTime:DescribeModelResponse', describeModelResponse_lastUpdatedTime - Indicates the last time the ML model was updated. The type of update is not specified.

$sel:trainingDataEndTime:DescribeModelResponse', describeModelResponse_trainingDataEndTime - Indicates the time reference in the dataset that was used to end the subset of training data for the ML model.

$sel:createdAt:DescribeModelResponse', describeModelResponse_createdAt - Indicates the time and date at which the ML model was created.

$sel:modelName:DescribeModel', describeModelResponse_modelName - The name of the ML model being described.

$sel:modelMetrics:DescribeModelResponse', describeModelResponse_modelMetrics - The Model Metrics show an aggregated summary of the model's performance within the evaluation time range. This is the JSON content of the metrics created when evaluating the model.

$sel:evaluationDataStartTime:DescribeModelResponse', describeModelResponse_evaluationDataStartTime - Indicates the time reference in the dataset that was used to begin the subset of evaluation data for the ML model.

$sel:schema:DescribeModelResponse', describeModelResponse_schema - A JSON description of the data that is in each time series dataset, including names, column names, and data types.

$sel:offCondition:DescribeModelResponse', describeModelResponse_offCondition - Indicates that the asset associated with this sensor has been shut off. As long as this condition is met, Lookout for Equipment will not use data from this asset for training, evaluation, or inference.

$sel:evaluationDataEndTime:DescribeModelResponse', describeModelResponse_evaluationDataEndTime - Indicates the time reference in the dataset that was used to end the subset of evaluation data for the ML model.

$sel:datasetName:DescribeModelResponse', describeModelResponse_datasetName - The name of the dataset being used by the ML being described.

$sel:trainingDataStartTime:DescribeModelResponse', describeModelResponse_trainingDataStartTime - Indicates the time reference in the dataset that was used to begin the subset of training data for the ML model.

$sel:trainingExecutionEndTime:DescribeModelResponse', describeModelResponse_trainingExecutionEndTime - Indicates the time at which the training of the ML model was completed.

$sel:labelsInputConfiguration:DescribeModelResponse', describeModelResponse_labelsInputConfiguration - Specifies configuration information about the labels input, including its S3 location.

$sel:serverSideKmsKeyId:DescribeModelResponse', describeModelResponse_serverSideKmsKeyId - Provides the identifier of the KMS key used to encrypt model data by Amazon Lookout for Equipment.

$sel:roleArn:DescribeModelResponse', describeModelResponse_roleArn - The Amazon Resource Name (ARN) of a role with permission to access the data source for the ML model being described.

$sel:httpStatus:DescribeModelResponse', describeModelResponse_httpStatus - The response's http status code.

Response Lenses

describeModelResponse_status :: Lens' DescribeModelResponse (Maybe ModelStatus)

Specifies the current status of the model being described. Status describes the status of the most recent action of the model.

describeModelResponse_dataPreProcessingConfiguration :: Lens' DescribeModelResponse (Maybe DataPreProcessingConfiguration)

The configuration is the TargetSamplingRate, which is the sampling rate of the data after post processing by Amazon Lookout for Equipment. For example, if you provide data that has been collected at a 1 second level and you want the system to resample the data at a 1 minute rate before training, the TargetSamplingRate is 1 minute.

When providing a value for the TargetSamplingRate, you must attach the prefix "PT" to the rate you want. The value for a 1 second rate is therefore PT1S, the value for a 15 minute rate is PT15M, and the value for a 1 hour rate is PT1H

describeModelResponse_trainingExecutionStartTime :: Lens' DescribeModelResponse (Maybe UTCTime)

Indicates the time at which the training of the ML model began.

describeModelResponse_datasetArn :: Lens' DescribeModelResponse (Maybe Text)

The Amazon Resouce Name (ARN) of the dataset used to create the ML model being described.

describeModelResponse_failedReason :: Lens' DescribeModelResponse (Maybe Text)

If the training of the ML model failed, this indicates the reason for that failure.

describeModelResponse_modelArn :: Lens' DescribeModelResponse (Maybe Text)

The Amazon Resource Name (ARN) of the ML model being described.

describeModelResponse_lastUpdatedTime :: Lens' DescribeModelResponse (Maybe UTCTime)

Indicates the last time the ML model was updated. The type of update is not specified.

describeModelResponse_trainingDataEndTime :: Lens' DescribeModelResponse (Maybe UTCTime)

Indicates the time reference in the dataset that was used to end the subset of training data for the ML model.

describeModelResponse_createdAt :: Lens' DescribeModelResponse (Maybe UTCTime)

Indicates the time and date at which the ML model was created.

describeModelResponse_modelName :: Lens' DescribeModelResponse (Maybe Text)

The name of the ML model being described.

describeModelResponse_modelMetrics :: Lens' DescribeModelResponse (Maybe Text)

The Model Metrics show an aggregated summary of the model's performance within the evaluation time range. This is the JSON content of the metrics created when evaluating the model.

describeModelResponse_evaluationDataStartTime :: Lens' DescribeModelResponse (Maybe UTCTime)

Indicates the time reference in the dataset that was used to begin the subset of evaluation data for the ML model.

describeModelResponse_schema :: Lens' DescribeModelResponse (Maybe Text)

A JSON description of the data that is in each time series dataset, including names, column names, and data types.

describeModelResponse_offCondition :: Lens' DescribeModelResponse (Maybe Text)

Indicates that the asset associated with this sensor has been shut off. As long as this condition is met, Lookout for Equipment will not use data from this asset for training, evaluation, or inference.

describeModelResponse_evaluationDataEndTime :: Lens' DescribeModelResponse (Maybe UTCTime)

Indicates the time reference in the dataset that was used to end the subset of evaluation data for the ML model.

describeModelResponse_datasetName :: Lens' DescribeModelResponse (Maybe Text)

The name of the dataset being used by the ML being described.

describeModelResponse_trainingDataStartTime :: Lens' DescribeModelResponse (Maybe UTCTime)

Indicates the time reference in the dataset that was used to begin the subset of training data for the ML model.

describeModelResponse_trainingExecutionEndTime :: Lens' DescribeModelResponse (Maybe UTCTime)

Indicates the time at which the training of the ML model was completed.

describeModelResponse_labelsInputConfiguration :: Lens' DescribeModelResponse (Maybe LabelsInputConfiguration)

Specifies configuration information about the labels input, including its S3 location.

describeModelResponse_serverSideKmsKeyId :: Lens' DescribeModelResponse (Maybe Text)

Provides the identifier of the KMS key used to encrypt model data by Amazon Lookout for Equipment.

describeModelResponse_roleArn :: Lens' DescribeModelResponse (Maybe Text)

The Amazon Resource Name (ARN) of a role with permission to access the data source for the ML model being described.

describeModelResponse_httpStatus :: Lens' DescribeModelResponse Int

The response's http status code.