Network.AWS.LookoutEquipment.StartDataIngestionJob

Description

Starts a data ingestion job. Amazon Lookout for Equipment returns the job status.

Creating a Request

data StartDataIngestionJob

See: newStartDataIngestionJob smart constructor.

Constructors

StartDataIngestionJob'

Fields

datasetName :: Text

The name of the dataset being used by the data ingestion job.

ingestionInputConfiguration :: IngestionInputConfiguration

Specifies information for the input data for the data ingestion job, including dataset S3 location.

roleArn :: Text

The Amazon Resource Name (ARN) of a role with permission to access the data source for the data ingestion job.

clientToken :: Text

A unique identifier for the request. If you do not set the client request token, Amazon Lookout for Equipment generates one.

newStartDataIngestionJob

Arguments

:: Text
$sel:datasetName:StartDataIngestionJob'

-> IngestionInputConfiguration
$sel:ingestionInputConfiguration:StartDataIngestionJob'

-> Text
$sel:roleArn:StartDataIngestionJob'

-> Text
$sel:clientToken:StartDataIngestionJob'

-> StartDataIngestionJob

Create a value of StartDataIngestionJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetName:StartDataIngestionJob', startDataIngestionJob_datasetName - The name of the dataset being used by the data ingestion job.

$sel:ingestionInputConfiguration:StartDataIngestionJob', startDataIngestionJob_ingestionInputConfiguration - Specifies information for the input data for the data ingestion job, including dataset S3 location.

$sel:roleArn:StartDataIngestionJob', startDataIngestionJob_roleArn - The Amazon Resource Name (ARN) of a role with permission to access the data source for the data ingestion job.

$sel:clientToken:StartDataIngestionJob', startDataIngestionJob_clientToken - A unique identifier for the request. If you do not set the client request token, Amazon Lookout for Equipment generates one.

Request Lenses

startDataIngestionJob_datasetName :: Lens' StartDataIngestionJob Text

The name of the dataset being used by the data ingestion job.

startDataIngestionJob_ingestionInputConfiguration :: Lens' StartDataIngestionJob IngestionInputConfiguration

Specifies information for the input data for the data ingestion job, including dataset S3 location.

startDataIngestionJob_roleArn :: Lens' StartDataIngestionJob Text

The Amazon Resource Name (ARN) of a role with permission to access the data source for the data ingestion job.

startDataIngestionJob_clientToken :: Lens' StartDataIngestionJob Text

A unique identifier for the request. If you do not set the client request token, Amazon Lookout for Equipment generates one.

Destructuring the Response

data StartDataIngestionJobResponse

See: newStartDataIngestionJobResponse smart constructor.

Constructors

StartDataIngestionJobResponse'

Fields

status :: Maybe IngestionJobStatus

Indicates the status of the StartDataIngestionJob operation.

jobId :: Maybe Text

Indicates the job ID of the data ingestion job.

httpStatus :: Int

The response's http status code.

newStartDataIngestionJobResponse

Arguments

:: Int
$sel:httpStatus:StartDataIngestionJobResponse'

-> StartDataIngestionJobResponse

Create a value of StartDataIngestionJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:StartDataIngestionJobResponse', startDataIngestionJobResponse_status - Indicates the status of the StartDataIngestionJob operation.

$sel:jobId:StartDataIngestionJobResponse', startDataIngestionJobResponse_jobId - Indicates the job ID of the data ingestion job.

$sel:httpStatus:StartDataIngestionJobResponse', startDataIngestionJobResponse_httpStatus - The response's http status code.

Response Lenses

startDataIngestionJobResponse_status :: Lens' StartDataIngestionJobResponse (Maybe IngestionJobStatus)

Indicates the status of the StartDataIngestionJob operation.

startDataIngestionJobResponse_jobId :: Lens' StartDataIngestionJobResponse (Maybe Text)

Indicates the job ID of the data ingestion job.

startDataIngestionJobResponse_httpStatus :: Lens' StartDataIngestionJobResponse Int

The response's http status code.