Network.AWS.LookoutEquipment.Types.DatasetSummary

Description

Documentation

data DatasetSummary

Contains information about the specific data set, including name, ARN, and status.

See: newDatasetSummary smart constructor.

Constructors

DatasetSummary'

Fields

status :: Maybe DatasetStatus

Indicates the status of the dataset.

datasetArn :: Maybe Text

The Amazon Resource Name (ARN) of the specified dataset.

createdAt :: Maybe POSIX

The time at which the dataset was created in Amazon Lookout for Equipment.

datasetName :: Maybe Text

The name of the dataset.

newDatasetSummary :: DatasetSummary

Create a value of DatasetSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DatasetSummary', datasetSummary_status - Indicates the status of the dataset.

$sel:datasetArn:DatasetSummary', datasetSummary_datasetArn - The Amazon Resource Name (ARN) of the specified dataset.

$sel:createdAt:DatasetSummary', datasetSummary_createdAt - The time at which the dataset was created in Amazon Lookout for Equipment.

$sel:datasetName:DatasetSummary', datasetSummary_datasetName - The name of the dataset.

datasetSummary_status :: Lens' DatasetSummary (Maybe DatasetStatus)

Indicates the status of the dataset.

datasetSummary_datasetArn :: Lens' DatasetSummary (Maybe Text)

The Amazon Resource Name (ARN) of the specified dataset.

datasetSummary_createdAt :: Lens' DatasetSummary (Maybe UTCTime)

The time at which the dataset was created in Amazon Lookout for Equipment.

datasetSummary_datasetName :: Lens' DatasetSummary (Maybe Text)

The name of the dataset.