Network.AWS.LookoutEquipment.Types.InferenceExecutionSummary

Description

Documentation

data InferenceExecutionSummary

Contains information about the specific inference execution, including input and output data configuration, inference scheduling information, status, and so on.

See: newInferenceExecutionSummary smart constructor.

Constructors

InferenceExecutionSummary'

Fields

status :: Maybe InferenceExecutionStatus

Indicates the status of the inference execution.

failedReason :: Maybe Text

Specifies the reason for failure when an inference execution has failed.

modelArn :: Maybe Text

The Amazon Resource Name (ARN) of the ML model used for the inference execution.

dataStartTime :: Maybe POSIX

Indicates the time reference in the dataset at which the inference execution began.

modelName :: Maybe Text

The name of the ML model being used for the inference execution.

customerResultObject :: Maybe S3Object

inferenceSchedulerArn :: Maybe Text

The Amazon Resource Name (ARN) of the inference scheduler being used for the inference execution.

scheduledStartTime :: Maybe POSIX

Indicates the start time at which the inference scheduler began the specific inference execution.

dataOutputConfiguration :: Maybe InferenceOutputConfiguration

Specifies configuration information for the output results from for the inference execution, including the output S3 location.

dataEndTime :: Maybe POSIX

Indicates the time reference in the dataset at which the inference execution stopped.

inferenceSchedulerName :: Maybe Text

The name of the inference scheduler being used for the inference execution.

dataInputConfiguration :: Maybe InferenceInputConfiguration

Specifies configuration information for the input data for the inference scheduler, including delimiter, format, and dataset location.

newInferenceExecutionSummary :: InferenceExecutionSummary

Create a value of InferenceExecutionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:InferenceExecutionSummary', inferenceExecutionSummary_status - Indicates the status of the inference execution.

$sel:failedReason:InferenceExecutionSummary', inferenceExecutionSummary_failedReason - Specifies the reason for failure when an inference execution has failed.

$sel:modelArn:InferenceExecutionSummary', inferenceExecutionSummary_modelArn - The Amazon Resource Name (ARN) of the ML model used for the inference execution.

$sel:dataStartTime:InferenceExecutionSummary', inferenceExecutionSummary_dataStartTime - Indicates the time reference in the dataset at which the inference execution began.

$sel:modelName:InferenceExecutionSummary', inferenceExecutionSummary_modelName - The name of the ML model being used for the inference execution.

$sel:customerResultObject:InferenceExecutionSummary', inferenceExecutionSummary_customerResultObject -

$sel:inferenceSchedulerArn:InferenceExecutionSummary', inferenceExecutionSummary_inferenceSchedulerArn - The Amazon Resource Name (ARN) of the inference scheduler being used for the inference execution.

$sel:scheduledStartTime:InferenceExecutionSummary', inferenceExecutionSummary_scheduledStartTime - Indicates the start time at which the inference scheduler began the specific inference execution.

$sel:dataOutputConfiguration:InferenceExecutionSummary', inferenceExecutionSummary_dataOutputConfiguration - Specifies configuration information for the output results from for the inference execution, including the output S3 location.

$sel:dataEndTime:InferenceExecutionSummary', inferenceExecutionSummary_dataEndTime - Indicates the time reference in the dataset at which the inference execution stopped.

$sel:inferenceSchedulerName:InferenceExecutionSummary', inferenceExecutionSummary_inferenceSchedulerName - The name of the inference scheduler being used for the inference execution.

$sel:dataInputConfiguration:InferenceExecutionSummary', inferenceExecutionSummary_dataInputConfiguration - Specifies configuration information for the input data for the inference scheduler, including delimiter, format, and dataset location.

inferenceExecutionSummary_status :: Lens' InferenceExecutionSummary (Maybe InferenceExecutionStatus)

Indicates the status of the inference execution.

inferenceExecutionSummary_failedReason :: Lens' InferenceExecutionSummary (Maybe Text)

Specifies the reason for failure when an inference execution has failed.

inferenceExecutionSummary_modelArn :: Lens' InferenceExecutionSummary (Maybe Text)

The Amazon Resource Name (ARN) of the ML model used for the inference execution.

inferenceExecutionSummary_dataStartTime :: Lens' InferenceExecutionSummary (Maybe UTCTime)

Indicates the time reference in the dataset at which the inference execution began.

inferenceExecutionSummary_modelName :: Lens' InferenceExecutionSummary (Maybe Text)

The name of the ML model being used for the inference execution.

inferenceExecutionSummary_customerResultObject :: Lens' InferenceExecutionSummary (Maybe S3Object)

inferenceExecutionSummary_inferenceSchedulerArn :: Lens' InferenceExecutionSummary (Maybe Text)

The Amazon Resource Name (ARN) of the inference scheduler being used for the inference execution.

inferenceExecutionSummary_scheduledStartTime :: Lens' InferenceExecutionSummary (Maybe UTCTime)

Indicates the start time at which the inference scheduler began the specific inference execution.

inferenceExecutionSummary_dataOutputConfiguration :: Lens' InferenceExecutionSummary (Maybe InferenceOutputConfiguration)

Specifies configuration information for the output results from for the inference execution, including the output S3 location.

inferenceExecutionSummary_dataEndTime :: Lens' InferenceExecutionSummary (Maybe UTCTime)

Indicates the time reference in the dataset at which the inference execution stopped.

inferenceExecutionSummary_inferenceSchedulerName :: Lens' InferenceExecutionSummary (Maybe Text)

The name of the inference scheduler being used for the inference execution.

inferenceExecutionSummary_dataInputConfiguration :: Lens' InferenceExecutionSummary (Maybe InferenceInputConfiguration)

Specifies configuration information for the input data for the inference scheduler, including delimiter, format, and dataset location.