Network.AWS.LookoutEquipment.UpdateInferenceScheduler

Description

Updates an inference scheduler.

Creating a Request

data UpdateInferenceScheduler

See: newUpdateInferenceScheduler smart constructor.

Constructors

UpdateInferenceScheduler'

Fields

dataUploadFrequency :: Maybe DataUploadFrequency

How often data is uploaded to the source S3 bucket for the input data. The value chosen is the length of time between data uploads. For instance, if you select 5 minutes, Amazon Lookout for Equipment will upload the real-time data to the source bucket once every 5 minutes. This frequency also determines how often Amazon Lookout for Equipment starts a scheduled inference on your data. In this example, it starts once every 5 minutes.

dataDelayOffsetInMinutes :: Maybe Natural

A period of time (in minutes) by which inference on the data is delayed after the data starts. For instance, if you select an offset delay time of five minutes, inference will not begin on the data until the first data measurement after the five minute mark. For example, if five minutes is selected, the inference scheduler will wake up at the configured frequency with the additional five minute delay time to check the customer S3 bucket. The customer can upload data at the same frequency and they don't need to stop and restart the scheduler when uploading new data.

dataOutputConfiguration :: Maybe InferenceOutputConfiguration

Specifies information for the output results from the inference scheduler, including the output S3 location.

dataInputConfiguration :: Maybe InferenceInputConfiguration

Specifies information for the input data for the inference scheduler, including delimiter, format, and dataset location.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of a role with permission to access the data source for the inference scheduler.

inferenceSchedulerName :: Text

The name of the inference scheduler to be updated.

newUpdateInferenceScheduler

Arguments

:: Text
$sel:inferenceSchedulerName:UpdateInferenceScheduler'

-> UpdateInferenceScheduler

Create a value of UpdateInferenceScheduler with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataUploadFrequency:UpdateInferenceScheduler', updateInferenceScheduler_dataUploadFrequency - How often data is uploaded to the source S3 bucket for the input data. The value chosen is the length of time between data uploads. For instance, if you select 5 minutes, Amazon Lookout for Equipment will upload the real-time data to the source bucket once every 5 minutes. This frequency also determines how often Amazon Lookout for Equipment starts a scheduled inference on your data. In this example, it starts once every 5 minutes.

$sel:dataDelayOffsetInMinutes:UpdateInferenceScheduler', updateInferenceScheduler_dataDelayOffsetInMinutes - A period of time (in minutes) by which inference on the data is delayed after the data starts. For instance, if you select an offset delay time of five minutes, inference will not begin on the data until the first data measurement after the five minute mark. For example, if five minutes is selected, the inference scheduler will wake up at the configured frequency with the additional five minute delay time to check the customer S3 bucket. The customer can upload data at the same frequency and they don't need to stop and restart the scheduler when uploading new data.

$sel:dataOutputConfiguration:UpdateInferenceScheduler', updateInferenceScheduler_dataOutputConfiguration - Specifies information for the output results from the inference scheduler, including the output S3 location.

$sel:dataInputConfiguration:UpdateInferenceScheduler', updateInferenceScheduler_dataInputConfiguration - Specifies information for the input data for the inference scheduler, including delimiter, format, and dataset location.

$sel:roleArn:UpdateInferenceScheduler', updateInferenceScheduler_roleArn - The Amazon Resource Name (ARN) of a role with permission to access the data source for the inference scheduler.

$sel:inferenceSchedulerName:UpdateInferenceScheduler', updateInferenceScheduler_inferenceSchedulerName - The name of the inference scheduler to be updated.

Request Lenses

updateInferenceScheduler_dataUploadFrequency :: Lens' UpdateInferenceScheduler (Maybe DataUploadFrequency)

How often data is uploaded to the source S3 bucket for the input data. The value chosen is the length of time between data uploads. For instance, if you select 5 minutes, Amazon Lookout for Equipment will upload the real-time data to the source bucket once every 5 minutes. This frequency also determines how often Amazon Lookout for Equipment starts a scheduled inference on your data. In this example, it starts once every 5 minutes.

updateInferenceScheduler_dataDelayOffsetInMinutes :: Lens' UpdateInferenceScheduler (Maybe Natural)

A period of time (in minutes) by which inference on the data is delayed after the data starts. For instance, if you select an offset delay time of five minutes, inference will not begin on the data until the first data measurement after the five minute mark. For example, if five minutes is selected, the inference scheduler will wake up at the configured frequency with the additional five minute delay time to check the customer S3 bucket. The customer can upload data at the same frequency and they don't need to stop and restart the scheduler when uploading new data.

updateInferenceScheduler_dataOutputConfiguration :: Lens' UpdateInferenceScheduler (Maybe InferenceOutputConfiguration)

Specifies information for the output results from the inference scheduler, including the output S3 location.

updateInferenceScheduler_dataInputConfiguration :: Lens' UpdateInferenceScheduler (Maybe InferenceInputConfiguration)

Specifies information for the input data for the inference scheduler, including delimiter, format, and dataset location.

updateInferenceScheduler_roleArn :: Lens' UpdateInferenceScheduler (Maybe Text)

The Amazon Resource Name (ARN) of a role with permission to access the data source for the inference scheduler.

updateInferenceScheduler_inferenceSchedulerName :: Lens' UpdateInferenceScheduler Text

The name of the inference scheduler to be updated.

Destructuring the Response

data UpdateInferenceSchedulerResponse

See: newUpdateInferenceSchedulerResponse smart constructor.

Constructors

UpdateInferenceSchedulerResponse'

newUpdateInferenceSchedulerResponse :: UpdateInferenceSchedulerResponse

Create a value of UpdateInferenceSchedulerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.