Network.AWS.LookoutMetrics.CreateAnomalyDetector

Description

Creates an anomaly detector.

Creating a Request

data CreateAnomalyDetector

See: newCreateAnomalyDetector smart constructor.

Constructors

CreateAnomalyDetector'

Fields

kmsKeyArn :: Maybe Text

The ARN of the KMS key to use to encrypt your data.

anomalyDetectorDescription :: Maybe Text

A description of the detector.

tags :: Maybe (HashMap Text Text)

A list of tags to apply to the anomaly detector.

anomalyDetectorName :: Text

The name of the detector.

anomalyDetectorConfig :: AnomalyDetectorConfig

Contains information about the configuration of the anomaly detector.

newCreateAnomalyDetector

Arguments

:: Text
$sel:anomalyDetectorName:CreateAnomalyDetector'

-> AnomalyDetectorConfig
$sel:anomalyDetectorConfig:CreateAnomalyDetector'

-> CreateAnomalyDetector

Create a value of CreateAnomalyDetector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyArn:CreateAnomalyDetector', createAnomalyDetector_kmsKeyArn - The ARN of the KMS key to use to encrypt your data.

$sel:anomalyDetectorDescription:CreateAnomalyDetector', createAnomalyDetector_anomalyDetectorDescription - A description of the detector.

$sel:tags:CreateAnomalyDetector', createAnomalyDetector_tags - A list of tags to apply to the anomaly detector.

$sel:anomalyDetectorName:CreateAnomalyDetector', createAnomalyDetector_anomalyDetectorName - The name of the detector.

$sel:anomalyDetectorConfig:CreateAnomalyDetector', createAnomalyDetector_anomalyDetectorConfig - Contains information about the configuration of the anomaly detector.

Request Lenses

createAnomalyDetector_kmsKeyArn :: Lens' CreateAnomalyDetector (Maybe Text)

The ARN of the KMS key to use to encrypt your data.

createAnomalyDetector_anomalyDetectorDescription :: Lens' CreateAnomalyDetector (Maybe Text)

A description of the detector.

createAnomalyDetector_tags :: Lens' CreateAnomalyDetector (Maybe (HashMap Text Text))

A list of tags to apply to the anomaly detector.

createAnomalyDetector_anomalyDetectorName :: Lens' CreateAnomalyDetector Text

The name of the detector.

createAnomalyDetector_anomalyDetectorConfig :: Lens' CreateAnomalyDetector AnomalyDetectorConfig

Contains information about the configuration of the anomaly detector.

Destructuring the Response

data CreateAnomalyDetectorResponse

See: newCreateAnomalyDetectorResponse smart constructor.

Constructors

CreateAnomalyDetectorResponse'

Fields

anomalyDetectorArn :: Maybe Text

The ARN of the detector.

httpStatus :: Int

The response's http status code.

newCreateAnomalyDetectorResponse

Arguments

:: Int
$sel:httpStatus:CreateAnomalyDetectorResponse'

-> CreateAnomalyDetectorResponse

Create a value of CreateAnomalyDetectorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:anomalyDetectorArn:CreateAnomalyDetectorResponse', createAnomalyDetectorResponse_anomalyDetectorArn - The ARN of the detector.

$sel:httpStatus:CreateAnomalyDetectorResponse', createAnomalyDetectorResponse_httpStatus - The response's http status code.

Response Lenses

createAnomalyDetectorResponse_anomalyDetectorArn :: Lens' CreateAnomalyDetectorResponse (Maybe Text)

The ARN of the detector.

createAnomalyDetectorResponse_httpStatus :: Lens' CreateAnomalyDetectorResponse Int

The response's http status code.