Network.AWS.LookoutMetrics.DescribeMetricSet

Description

Describes a dataset.

Amazon Lookout for Metrics API actions are eventually consistent. If you do a read operation on a resource immediately after creating or modifying it, use retries to allow time for the write operation to complete.

Creating a Request

data DescribeMetricSet

See: newDescribeMetricSet smart constructor.

Constructors

DescribeMetricSet'

Fields

metricSetArn :: Text

The ARN of the dataset.

newDescribeMetricSet

Arguments

:: Text
$sel:metricSetArn:DescribeMetricSet'

-> DescribeMetricSet

Create a value of DescribeMetricSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricSetArn:DescribeMetricSet', describeMetricSet_metricSetArn - The ARN of the dataset.

Request Lenses

describeMetricSet_metricSetArn :: Lens' DescribeMetricSet Text

The ARN of the dataset.

Destructuring the Response

data DescribeMetricSetResponse

See: newDescribeMetricSetResponse smart constructor.

Constructors

DescribeMetricSetResponse'

Fields

creationTime :: Maybe POSIX

The time at which the dataset was created.

dimensionList :: Maybe (NonEmpty Text)

A list of the dimensions chosen for analysis.

offset :: Maybe Natural

The offset in seconds. Only supported for S3 and Redshift datasources.

timestampColumn :: Maybe TimestampColumn

Contains information about the column used for tracking time in your source data.

metricList :: Maybe (NonEmpty Metric)

A list of the metrics defined by the dataset.

anomalyDetectorArn :: Maybe Text

The ARN of the detector that contains the dataset.

metricSource :: Maybe MetricSource

Contains information about the dataset's source data.

metricSetName :: Maybe Text

The name of the dataset.

metricSetFrequency :: Maybe Frequency

The interval at which the data will be analyzed for anomalies.

metricSetDescription :: Maybe Text

The dataset's description.

timezone :: Maybe Text

The time zone in which the dataset's data was recorded.

metricSetArn :: Maybe Text

The ARN of the dataset.

lastModificationTime :: Maybe POSIX

The time at which the dataset was last modified.

httpStatus :: Int

The response's http status code.

newDescribeMetricSetResponse

Arguments

:: Int
$sel:httpStatus:DescribeMetricSetResponse'

-> DescribeMetricSetResponse

Create a value of DescribeMetricSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeMetricSetResponse', describeMetricSetResponse_creationTime - The time at which the dataset was created.

$sel:dimensionList:DescribeMetricSetResponse', describeMetricSetResponse_dimensionList - A list of the dimensions chosen for analysis.

$sel:offset:DescribeMetricSetResponse', describeMetricSetResponse_offset - The offset in seconds. Only supported for S3 and Redshift datasources.

$sel:timestampColumn:DescribeMetricSetResponse', describeMetricSetResponse_timestampColumn - Contains information about the column used for tracking time in your source data.

$sel:metricList:DescribeMetricSetResponse', describeMetricSetResponse_metricList - A list of the metrics defined by the dataset.

$sel:anomalyDetectorArn:DescribeMetricSetResponse', describeMetricSetResponse_anomalyDetectorArn - The ARN of the detector that contains the dataset.

$sel:metricSource:DescribeMetricSetResponse', describeMetricSetResponse_metricSource - Contains information about the dataset's source data.

$sel:metricSetName:DescribeMetricSetResponse', describeMetricSetResponse_metricSetName - The name of the dataset.

$sel:metricSetFrequency:DescribeMetricSetResponse', describeMetricSetResponse_metricSetFrequency - The interval at which the data will be analyzed for anomalies.

$sel:metricSetDescription:DescribeMetricSetResponse', describeMetricSetResponse_metricSetDescription - The dataset's description.

$sel:timezone:DescribeMetricSetResponse', describeMetricSetResponse_timezone - The time zone in which the dataset's data was recorded.

$sel:metricSetArn:DescribeMetricSet', describeMetricSetResponse_metricSetArn - The ARN of the dataset.

$sel:lastModificationTime:DescribeMetricSetResponse', describeMetricSetResponse_lastModificationTime - The time at which the dataset was last modified.

$sel:httpStatus:DescribeMetricSetResponse', describeMetricSetResponse_httpStatus - The response's http status code.

Response Lenses

describeMetricSetResponse_creationTime :: Lens' DescribeMetricSetResponse (Maybe UTCTime)

The time at which the dataset was created.

describeMetricSetResponse_dimensionList :: Lens' DescribeMetricSetResponse (Maybe (NonEmpty Text))

A list of the dimensions chosen for analysis.

describeMetricSetResponse_offset :: Lens' DescribeMetricSetResponse (Maybe Natural)

The offset in seconds. Only supported for S3 and Redshift datasources.

describeMetricSetResponse_timestampColumn :: Lens' DescribeMetricSetResponse (Maybe TimestampColumn)

Contains information about the column used for tracking time in your source data.

describeMetricSetResponse_metricList :: Lens' DescribeMetricSetResponse (Maybe (NonEmpty Metric))

A list of the metrics defined by the dataset.

describeMetricSetResponse_anomalyDetectorArn :: Lens' DescribeMetricSetResponse (Maybe Text)

The ARN of the detector that contains the dataset.

describeMetricSetResponse_metricSource :: Lens' DescribeMetricSetResponse (Maybe MetricSource)

Contains information about the dataset's source data.

describeMetricSetResponse_metricSetName :: Lens' DescribeMetricSetResponse (Maybe Text)

The name of the dataset.

describeMetricSetResponse_metricSetFrequency :: Lens' DescribeMetricSetResponse (Maybe Frequency)

The interval at which the data will be analyzed for anomalies.

describeMetricSetResponse_metricSetDescription :: Lens' DescribeMetricSetResponse (Maybe Text)

The dataset's description.

describeMetricSetResponse_timezone :: Lens' DescribeMetricSetResponse (Maybe Text)

The time zone in which the dataset's data was recorded.

describeMetricSetResponse_metricSetArn :: Lens' DescribeMetricSetResponse (Maybe Text)

The ARN of the dataset.

describeMetricSetResponse_lastModificationTime :: Lens' DescribeMetricSetResponse (Maybe UTCTime)

The time at which the dataset was last modified.

describeMetricSetResponse_httpStatus :: Lens' DescribeMetricSetResponse Int

The response's http status code.