Network.AWS.LookoutMetrics.GetAnomalyGroup

Description

Returns details about a group of anomalous metrics.

Creating a Request

data GetAnomalyGroup

See: newGetAnomalyGroup smart constructor.

Constructors

GetAnomalyGroup'

Fields

anomalyGroupId :: Text

The ID of the anomaly group.

anomalyDetectorArn :: Text

The Amazon Resource Name (ARN) of the anomaly detector.

newGetAnomalyGroup

Arguments

:: Text
$sel:anomalyGroupId:GetAnomalyGroup'

-> Text
$sel:anomalyDetectorArn:GetAnomalyGroup'

-> GetAnomalyGroup

Create a value of GetAnomalyGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:anomalyGroupId:GetAnomalyGroup', getAnomalyGroup_anomalyGroupId - The ID of the anomaly group.

$sel:anomalyDetectorArn:GetAnomalyGroup', getAnomalyGroup_anomalyDetectorArn - The Amazon Resource Name (ARN) of the anomaly detector.

Request Lenses

getAnomalyGroup_anomalyGroupId :: Lens' GetAnomalyGroup Text

The ID of the anomaly group.

getAnomalyGroup_anomalyDetectorArn :: Lens' GetAnomalyGroup Text

The Amazon Resource Name (ARN) of the anomaly detector.

Destructuring the Response

data GetAnomalyGroupResponse

See: newGetAnomalyGroupResponse smart constructor.

Constructors

GetAnomalyGroupResponse'

Fields

anomalyGroup :: Maybe AnomalyGroup

Details about the anomaly group.

httpStatus :: Int

The response's http status code.

newGetAnomalyGroupResponse

Arguments

:: Int
$sel:httpStatus:GetAnomalyGroupResponse'

-> GetAnomalyGroupResponse

Create a value of GetAnomalyGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:anomalyGroup:GetAnomalyGroupResponse', getAnomalyGroupResponse_anomalyGroup - Details about the anomaly group.

$sel:httpStatus:GetAnomalyGroupResponse', getAnomalyGroupResponse_httpStatus - The response's http status code.

Response Lenses

getAnomalyGroupResponse_anomalyGroup :: Lens' GetAnomalyGroupResponse (Maybe AnomalyGroup)

Details about the anomaly group.

getAnomalyGroupResponse_httpStatus :: Lens' GetAnomalyGroupResponse Int

The response's http status code.