Network.AWS.LookoutMetrics.Lens

Description

Operations

GetFeedback

getFeedback_nextToken :: Lens' GetFeedback (Maybe Text)

Specify the pagination token that's returned by a previous request to retrieve the next page of results.

getFeedback_maxResults :: Lens' GetFeedback (Maybe Natural)

The maximum number of results to return.

getFeedback_anomalyDetectorArn :: Lens' GetFeedback Text

The Amazon Resource Name (ARN) of the anomaly detector.

getFeedback_anomalyGroupTimeSeriesFeedback :: Lens' GetFeedback AnomalyGroupTimeSeries

The anomalous metric and group ID.

getFeedbackResponse_anomalyGroupTimeSeriesFeedback :: Lens' GetFeedbackResponse (Maybe [TimeSeriesFeedback])

Feedback for an anomalous metric.

getFeedbackResponse_nextToken :: Lens' GetFeedbackResponse (Maybe Text)

The pagination token that's included if more results are available.

getFeedbackResponse_httpStatus :: Lens' GetFeedbackResponse Int

The response's http status code.

ListAlerts

listAlerts_anomalyDetectorArn :: Lens' ListAlerts (Maybe Text)

The ARN of the alert's detector.

listAlerts_nextToken :: Lens' ListAlerts (Maybe Text)

If the result of the previous request is truncated, the response includes a NextToken. To retrieve the next set of results, use the token in the next request. Tokens expire after 24 hours.

listAlerts_maxResults :: Lens' ListAlerts (Maybe Natural)

The maximum number of results that will be displayed by the request.

listAlertsResponse_nextToken :: Lens' ListAlertsResponse (Maybe Text)

If the response is truncated, the service returns this token. To retrieve the next set of results, use this token in the next request.

listAlertsResponse_alertSummaryList :: Lens' ListAlertsResponse (Maybe [AlertSummary])

Contains information about an alert.

listAlertsResponse_httpStatus :: Lens' ListAlertsResponse Int

The response's http status code.

ListMetricSets

listMetricSets_anomalyDetectorArn :: Lens' ListMetricSets (Maybe Text)

The ARN of the anomaly detector containing the metrics sets to list.

listMetricSets_nextToken :: Lens' ListMetricSets (Maybe Text)

If the result of the previous request was truncated, the response includes a NextToken. To retrieve the next set of results, use the token in the next request. Tokens expire after 24 hours.

listMetricSets_maxResults :: Lens' ListMetricSets (Maybe Natural)

The maximum number of results to return.

listMetricSetsResponse_nextToken :: Lens' ListMetricSetsResponse (Maybe Text)

If the response is truncated, the list call returns this token. To retrieve the next set of results, use the token in the next list request.

listMetricSetsResponse_metricSetSummaryList :: Lens' ListMetricSetsResponse (Maybe [MetricSetSummary])

A list of the datasets in the AWS Region, with configuration details for each.

listMetricSetsResponse_httpStatus :: Lens' ListMetricSetsResponse Int

The response's http status code.

DeleteAnomalyDetector

deleteAnomalyDetector_anomalyDetectorArn :: Lens' DeleteAnomalyDetector Text

The ARN of the detector to delete.

deleteAnomalyDetectorResponse_httpStatus :: Lens' DeleteAnomalyDetectorResponse Int

The response's http status code.

UpdateAnomalyDetector

updateAnomalyDetector_kmsKeyArn :: Lens' UpdateAnomalyDetector (Maybe Text)

The Amazon Resource Name (ARN) of an AWS KMS encryption key.

updateAnomalyDetector_anomalyDetectorConfig :: Lens' UpdateAnomalyDetector (Maybe AnomalyDetectorConfig)

Contains information about the configuration to which the detector will be updated.

updateAnomalyDetector_anomalyDetectorDescription :: Lens' UpdateAnomalyDetector (Maybe Text)

The updated detector description.

updateAnomalyDetector_anomalyDetectorArn :: Lens' UpdateAnomalyDetector Text

The ARN of the detector to update.

updateAnomalyDetectorResponse_anomalyDetectorArn :: Lens' UpdateAnomalyDetectorResponse (Maybe Text)

The ARN of the updated detector.

updateAnomalyDetectorResponse_httpStatus :: Lens' UpdateAnomalyDetectorResponse Int

The response's http status code.

ListAnomalyDetectors

listAnomalyDetectors_nextToken :: Lens' ListAnomalyDetectors (Maybe Text)

If the result of the previous request was truncated, the response includes a NextToken. To retrieve the next set of results, use the token in the next request. Tokens expire after 24 hours.

listAnomalyDetectors_maxResults :: Lens' ListAnomalyDetectors (Maybe Natural)

The maximum number of results to return.

listAnomalyDetectorsResponse_nextToken :: Lens' ListAnomalyDetectorsResponse (Maybe Text)

If the response is truncated, the service returns this token. To retrieve the next set of results, use the token in the next request.

listAnomalyDetectorsResponse_anomalyDetectorSummaryList :: Lens' ListAnomalyDetectorsResponse (Maybe [AnomalyDetectorSummary])

A list of anomaly detectors in the account in the current region.

listAnomalyDetectorsResponse_httpStatus :: Lens' ListAnomalyDetectorsResponse Int

The response's http status code.

DescribeAnomalyDetectionExecutions

describeAnomalyDetectionExecutions_nextToken :: Lens' DescribeAnomalyDetectionExecutions (Maybe Text)

Specify the pagination token that's returned by a previous request to retrieve the next page of results.

describeAnomalyDetectionExecutions_timestamp :: Lens' DescribeAnomalyDetectionExecutions (Maybe Text)

The timestamp of the anomaly detection job.

describeAnomalyDetectionExecutions_maxResults :: Lens' DescribeAnomalyDetectionExecutions (Maybe Natural)

The number of items to return in the response.

describeAnomalyDetectionExecutions_anomalyDetectorArn :: Lens' DescribeAnomalyDetectionExecutions Text

The Amazon Resource Name (ARN) of the anomaly detector.

describeAnomalyDetectionExecutionsResponse_nextToken :: Lens' DescribeAnomalyDetectionExecutionsResponse (Maybe Text)

The pagination token that's included if more results are available.

describeAnomalyDetectionExecutionsResponse_executionList :: Lens' DescribeAnomalyDetectionExecutionsResponse (Maybe [ExecutionStatus])

A list of detection jobs.

describeAnomalyDetectionExecutionsResponse_httpStatus :: Lens' DescribeAnomalyDetectionExecutionsResponse Int

The response's http status code.

CreateMetricSet

createMetricSet_dimensionList :: Lens' CreateMetricSet (Maybe (NonEmpty Text))

A list of the fields you want to treat as dimensions.

createMetricSet_offset :: Lens' CreateMetricSet (Maybe Natural)

After an interval ends, the amount of seconds that the detector waits before importing data. Offset is only supported for S3 and Redshift datasources.

createMetricSet_timestampColumn :: Lens' CreateMetricSet (Maybe TimestampColumn)

Contains information about the column used for tracking time in your source data.

createMetricSet_metricSetFrequency :: Lens' CreateMetricSet (Maybe Frequency)

The frequency with which the source data will be analyzed for anomalies.

createMetricSet_metricSetDescription :: Lens' CreateMetricSet (Maybe Text)

A description of the dataset you are creating.

createMetricSet_timezone :: Lens' CreateMetricSet (Maybe Text)

The time zone in which your source data was recorded.

createMetricSet_tags :: Lens' CreateMetricSet (Maybe (HashMap Text Text))

A list of tags to apply to the dataset.

createMetricSet_anomalyDetectorArn :: Lens' CreateMetricSet Text

The ARN of the anomaly detector that will use the dataset.

createMetricSet_metricSetName :: Lens' CreateMetricSet Text

The name of the dataset.

createMetricSet_metricList :: Lens' CreateMetricSet (NonEmpty Metric)

A list of metrics that the dataset will contain.

createMetricSet_metricSource :: Lens' CreateMetricSet MetricSource

Contains information about how the source data should be interpreted.

createMetricSetResponse_metricSetArn :: Lens' CreateMetricSetResponse (Maybe Text)

The ARN of the dataset.

createMetricSetResponse_httpStatus :: Lens' CreateMetricSetResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The resource's Amazon Resource Name (ARN).

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The resource's tags.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

CreateAlert

createAlert_alertDescription :: Lens' CreateAlert (Maybe Text)

A description of the alert.

createAlert_tags :: Lens' CreateAlert (Maybe (HashMap Text Text))

A list of tags to apply to the alert.

createAlert_alertName :: Lens' CreateAlert Text

The name of the alert.

createAlert_alertSensitivityThreshold :: Lens' CreateAlert Natural

An integer from 0 to 100 specifying the alert sensitivity threshold.

createAlert_anomalyDetectorArn :: Lens' CreateAlert Text

The ARN of the detector to which the alert is attached.

createAlert_action :: Lens' CreateAlert Action

Action that will be triggered when there is an alert.

createAlertResponse_alertArn :: Lens' CreateAlertResponse (Maybe Text)

The ARN of the alert.

createAlertResponse_httpStatus :: Lens' CreateAlertResponse Int

The response's http status code.

GetAnomalyGroup

getAnomalyGroup_anomalyGroupId :: Lens' GetAnomalyGroup Text

The ID of the anomaly group.

getAnomalyGroup_anomalyDetectorArn :: Lens' GetAnomalyGroup Text

The Amazon Resource Name (ARN) of the anomaly detector.

getAnomalyGroupResponse_anomalyGroup :: Lens' GetAnomalyGroupResponse (Maybe AnomalyGroup)

Details about the anomaly group.

getAnomalyGroupResponse_httpStatus :: Lens' GetAnomalyGroupResponse Int

The response's http status code.

PutFeedback

putFeedback_anomalyDetectorArn :: Lens' PutFeedback Text

The Amazon Resource Name (ARN) of the anomaly detector.

putFeedback_anomalyGroupTimeSeriesFeedback :: Lens' PutFeedback AnomalyGroupTimeSeriesFeedback

Feedback for an anomalous metric.

putFeedbackResponse_httpStatus :: Lens' PutFeedbackResponse Int

The response's http status code.

BackTestAnomalyDetector

backTestAnomalyDetector_anomalyDetectorArn :: Lens' BackTestAnomalyDetector Text

The Amazon Resource Name (ARN) of the anomaly detector.

backTestAnomalyDetectorResponse_httpStatus :: Lens' BackTestAnomalyDetectorResponse Int

The response's http status code.

DeleteAlert

deleteAlert_alertArn :: Lens' DeleteAlert Text

The ARN of the alert to delete.

deleteAlertResponse_httpStatus :: Lens' DeleteAlertResponse Int

The response's http status code.

CreateAnomalyDetector

createAnomalyDetector_kmsKeyArn :: Lens' CreateAnomalyDetector (Maybe Text)

The ARN of the KMS key to use to encrypt your data.

createAnomalyDetector_anomalyDetectorDescription :: Lens' CreateAnomalyDetector (Maybe Text)

A description of the detector.

createAnomalyDetector_tags :: Lens' CreateAnomalyDetector (Maybe (HashMap Text Text))

A list of tags to apply to the anomaly detector.

createAnomalyDetector_anomalyDetectorName :: Lens' CreateAnomalyDetector Text

The name of the detector.

createAnomalyDetector_anomalyDetectorConfig :: Lens' CreateAnomalyDetector AnomalyDetectorConfig

Contains information about the configuration of the anomaly detector.

createAnomalyDetectorResponse_anomalyDetectorArn :: Lens' CreateAnomalyDetectorResponse (Maybe Text)

The ARN of the detector.

createAnomalyDetectorResponse_httpStatus :: Lens' CreateAnomalyDetectorResponse Int

The response's http status code.

UpdateMetricSet

updateMetricSet_dimensionList :: Lens' UpdateMetricSet (Maybe (NonEmpty Text))

The dimension list.

updateMetricSet_offset :: Lens' UpdateMetricSet (Maybe Natural)

After an interval ends, the amount of seconds that the detector waits before importing data. Offset is only supported for S3 and Redshift datasources.

updateMetricSet_timestampColumn :: Lens' UpdateMetricSet (Maybe TimestampColumn)

The timestamp column.

updateMetricSet_metricList :: Lens' UpdateMetricSet (Maybe (NonEmpty Metric))

The metric list.

updateMetricSet_metricSource :: Lens' UpdateMetricSet (Maybe MetricSource)

Undocumented member.

updateMetricSet_metricSetFrequency :: Lens' UpdateMetricSet (Maybe Frequency)

The dataset's interval.

updateMetricSet_metricSetDescription :: Lens' UpdateMetricSet (Maybe Text)

The dataset's description.

updateMetricSet_metricSetArn :: Lens' UpdateMetricSet Text

The ARN of the dataset to update.

updateMetricSetResponse_metricSetArn :: Lens' UpdateMetricSetResponse (Maybe Text)

The ARN of the dataset.

updateMetricSetResponse_httpStatus :: Lens' UpdateMetricSetResponse Int

The response's http status code.

ActivateAnomalyDetector

activateAnomalyDetector_anomalyDetectorArn :: Lens' ActivateAnomalyDetector Text

The ARN of the anomaly detector.

activateAnomalyDetectorResponse_httpStatus :: Lens' ActivateAnomalyDetectorResponse Int

The response's http status code.

ListAnomalyGroupTimeSeries

listAnomalyGroupTimeSeries_nextToken :: Lens' ListAnomalyGroupTimeSeries (Maybe Text)

Specify the pagination token that's returned by a previous request to retrieve the next page of results.

listAnomalyGroupTimeSeries_maxResults :: Lens' ListAnomalyGroupTimeSeries (Maybe Natural)

The maximum number of results to return.

listAnomalyGroupTimeSeries_anomalyDetectorArn :: Lens' ListAnomalyGroupTimeSeries Text

The Amazon Resource Name (ARN) of the anomaly detector.

listAnomalyGroupTimeSeries_anomalyGroupId :: Lens' ListAnomalyGroupTimeSeries Text

The ID of the anomaly group.

listAnomalyGroupTimeSeries_metricName :: Lens' ListAnomalyGroupTimeSeries Text

The name of the measure field.

listAnomalyGroupTimeSeriesResponse_timeSeriesList :: Lens' ListAnomalyGroupTimeSeriesResponse (Maybe [TimeSeries])

A list of anomalous metrics.

listAnomalyGroupTimeSeriesResponse_timestampList :: Lens' ListAnomalyGroupTimeSeriesResponse (Maybe [Text])

Timestamps for the anomalous metrics.

listAnomalyGroupTimeSeriesResponse_metricName :: Lens' ListAnomalyGroupTimeSeriesResponse (Maybe Text)

The name of the measure field.

listAnomalyGroupTimeSeriesResponse_anomalyGroupId :: Lens' ListAnomalyGroupTimeSeriesResponse (Maybe Text)

The ID of the anomaly group.

listAnomalyGroupTimeSeriesResponse_nextToken :: Lens' ListAnomalyGroupTimeSeriesResponse (Maybe Text)

The pagination token that's included if more results are available.

listAnomalyGroupTimeSeriesResponse_httpStatus :: Lens' ListAnomalyGroupTimeSeriesResponse Int

The response's http status code.

GetSampleData

getSampleData_s3SourceConfig :: Lens' GetSampleData (Maybe SampleDataS3SourceConfig)

A datasource bucket in Amazon S3.

getSampleDataResponse_sampleRows :: Lens' GetSampleDataResponse (Maybe [[Text]])

A list of records.

getSampleDataResponse_headerValues :: Lens' GetSampleDataResponse (Maybe [Text])

A list of header labels for the records.

getSampleDataResponse_httpStatus :: Lens' GetSampleDataResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The resource's Amazon Resource Name (ARN).

tagResource_tags :: Lens' TagResource (HashMap Text Text)

Tags to apply to the resource. Tag keys and values can contain letters, numbers, spaces, and the following symbols: _.:/=+@-

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

DescribeMetricSet

describeMetricSet_metricSetArn :: Lens' DescribeMetricSet Text

The ARN of the dataset.

describeMetricSetResponse_creationTime :: Lens' DescribeMetricSetResponse (Maybe UTCTime)

The time at which the dataset was created.

describeMetricSetResponse_dimensionList :: Lens' DescribeMetricSetResponse (Maybe (NonEmpty Text))

A list of the dimensions chosen for analysis.

describeMetricSetResponse_offset :: Lens' DescribeMetricSetResponse (Maybe Natural)

The offset in seconds. Only supported for S3 and Redshift datasources.

describeMetricSetResponse_timestampColumn :: Lens' DescribeMetricSetResponse (Maybe TimestampColumn)

Contains information about the column used for tracking time in your source data.

describeMetricSetResponse_metricList :: Lens' DescribeMetricSetResponse (Maybe (NonEmpty Metric))

A list of the metrics defined by the dataset.

describeMetricSetResponse_anomalyDetectorArn :: Lens' DescribeMetricSetResponse (Maybe Text)

The ARN of the detector that contains the dataset.

describeMetricSetResponse_metricSource :: Lens' DescribeMetricSetResponse (Maybe MetricSource)

Contains information about the dataset's source data.

describeMetricSetResponse_metricSetName :: Lens' DescribeMetricSetResponse (Maybe Text)

The name of the dataset.

describeMetricSetResponse_metricSetFrequency :: Lens' DescribeMetricSetResponse (Maybe Frequency)

The interval at which the data will be analyzed for anomalies.

describeMetricSetResponse_metricSetDescription :: Lens' DescribeMetricSetResponse (Maybe Text)

The dataset's description.

describeMetricSetResponse_timezone :: Lens' DescribeMetricSetResponse (Maybe Text)

The time zone in which the dataset's data was recorded.

describeMetricSetResponse_metricSetArn :: Lens' DescribeMetricSetResponse (Maybe Text)

The ARN of the dataset.

describeMetricSetResponse_lastModificationTime :: Lens' DescribeMetricSetResponse (Maybe UTCTime)

The time at which the dataset was last modified.

describeMetricSetResponse_httpStatus :: Lens' DescribeMetricSetResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The resource's Amazon Resource Name (ARN).

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

Keys to remove from the resource's tags.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

DescribeAlert

describeAlert_alertArn :: Lens' DescribeAlert Text

The ARN of the alert to describe.

describeAlertResponse_alert :: Lens' DescribeAlertResponse (Maybe Alert)

Contains information about an alert.

describeAlertResponse_httpStatus :: Lens' DescribeAlertResponse Int

The response's http status code.

ListAnomalyGroupSummaries

listAnomalyGroupSummaries_nextToken :: Lens' ListAnomalyGroupSummaries (Maybe Text)

Specify the pagination token that's returned by a previous request to retrieve the next page of results.

listAnomalyGroupSummaries_maxResults :: Lens' ListAnomalyGroupSummaries (Maybe Natural)

The maximum number of results to return.

listAnomalyGroupSummaries_anomalyDetectorArn :: Lens' ListAnomalyGroupSummaries Text

The Amazon Resource Name (ARN) of the anomaly detector.

listAnomalyGroupSummaries_sensitivityThreshold :: Lens' ListAnomalyGroupSummaries Natural

The minimum severity score for inclusion in the output.

listAnomalyGroupSummariesResponse_anomalyGroupStatistics :: Lens' ListAnomalyGroupSummariesResponse (Maybe AnomalyGroupStatistics)

Aggregated details about the anomaly groups.

listAnomalyGroupSummariesResponse_nextToken :: Lens' ListAnomalyGroupSummariesResponse (Maybe Text)

The pagination token that's included if more results are available.

listAnomalyGroupSummariesResponse_anomalyGroupSummaryList :: Lens' ListAnomalyGroupSummariesResponse (Maybe [AnomalyGroupSummary])

A list of anomaly group summaries.

listAnomalyGroupSummariesResponse_httpStatus :: Lens' ListAnomalyGroupSummariesResponse Int

The response's http status code.

DescribeAnomalyDetector

describeAnomalyDetector_anomalyDetectorArn :: Lens' DescribeAnomalyDetector Text

The ARN of the detector to describe.

describeAnomalyDetectorResponse_creationTime :: Lens' DescribeAnomalyDetectorResponse (Maybe UTCTime)

The time at which the detector was created.

describeAnomalyDetectorResponse_status :: Lens' DescribeAnomalyDetectorResponse (Maybe AnomalyDetectorStatus)

The status of the detector.

describeAnomalyDetectorResponse_failureReason :: Lens' DescribeAnomalyDetectorResponse (Maybe Text)

The reason that the detector failed, if any.

describeAnomalyDetectorResponse_kmsKeyArn :: Lens' DescribeAnomalyDetectorResponse (Maybe Text)

The ARN of the KMS key to use to encrypt your data.

describeAnomalyDetectorResponse_anomalyDetectorArn :: Lens' DescribeAnomalyDetectorResponse (Maybe Text)

The ARN of the detector.

describeAnomalyDetectorResponse_anomalyDetectorConfig :: Lens' DescribeAnomalyDetectorResponse (Maybe AnomalyDetectorConfigSummary)

Contains information about the detector's configuration.

describeAnomalyDetectorResponse_anomalyDetectorName :: Lens' DescribeAnomalyDetectorResponse (Maybe Text)

The name of the detector.

describeAnomalyDetectorResponse_anomalyDetectorDescription :: Lens' DescribeAnomalyDetectorResponse (Maybe Text)

A description of the detector.

describeAnomalyDetectorResponse_lastModificationTime :: Lens' DescribeAnomalyDetectorResponse (Maybe UTCTime)

The time at which the detector was last modified.

describeAnomalyDetectorResponse_httpStatus :: Lens' DescribeAnomalyDetectorResponse Int

The response's http status code.

Types

Action

action_lambdaConfiguration :: Lens' Action (Maybe LambdaConfiguration)

A configuration for an AWS Lambda channel.

action_sNSConfiguration :: Lens' Action (Maybe SNSConfiguration)

A configuration for an Amazon SNS channel.

Alert

alert_creationTime :: Lens' Alert (Maybe UTCTime)

The time at which the alert was created.

alert_action :: Lens' Alert (Maybe Action)

Action that will be triggered when there is an alert.

alert_anomalyDetectorArn :: Lens' Alert (Maybe Text)

The ARN of the detector to which the alert is attached.

alert_alertName :: Lens' Alert (Maybe Text)

The name of the alert.

alert_alertSensitivityThreshold :: Lens' Alert (Maybe Natural)

The minimum severity for an anomaly to trigger the alert.

alert_alertStatus :: Lens' Alert (Maybe AlertStatus)

The status of the alert.

alert_alertDescription :: Lens' Alert (Maybe Text)

A description of the alert.

alert_alertArn :: Lens' Alert (Maybe Text)

The ARN of the alert.

alert_alertType :: Lens' Alert (Maybe AlertType)

The type of the alert.

alert_lastModificationTime :: Lens' Alert (Maybe UTCTime)

The time at which the alert was last modified.

AlertSummary

alertSummary_creationTime :: Lens' AlertSummary (Maybe UTCTime)

The time at which the alert was created.

alertSummary_anomalyDetectorArn :: Lens' AlertSummary (Maybe Text)

The ARN of the detector to which the alert is attached.

alertSummary_alertName :: Lens' AlertSummary (Maybe Text)

The name of the alert.

alertSummary_alertSensitivityThreshold :: Lens' AlertSummary (Maybe Natural)

The minimum severity for an anomaly to trigger the alert.

alertSummary_alertStatus :: Lens' AlertSummary (Maybe AlertStatus)

The status of the alert.

alertSummary_alertArn :: Lens' AlertSummary (Maybe Text)

The ARN of the alert.

alertSummary_alertType :: Lens' AlertSummary (Maybe AlertType)

The type of the alert.

alertSummary_tags :: Lens' AlertSummary (Maybe (HashMap Text Text))

The alert's tags.

alertSummary_lastModificationTime :: Lens' AlertSummary (Maybe UTCTime)

The time at which the alert was last modified.

AnomalyDetectorConfig

anomalyDetectorConfig_anomalyDetectorFrequency :: Lens' AnomalyDetectorConfig (Maybe Frequency)

The frequency at which the detector analyzes its source data.

AnomalyDetectorConfigSummary

anomalyDetectorConfigSummary_anomalyDetectorFrequency :: Lens' AnomalyDetectorConfigSummary (Maybe Frequency)

The interval at which the detector analyzes its source data.

AnomalyDetectorSummary

anomalyDetectorSummary_creationTime :: Lens' AnomalyDetectorSummary (Maybe UTCTime)

The time at which the detector was created.

anomalyDetectorSummary_status :: Lens' AnomalyDetectorSummary (Maybe AnomalyDetectorStatus)

The status of detector.

anomalyDetectorSummary_anomalyDetectorArn :: Lens' AnomalyDetectorSummary (Maybe Text)

The ARN of the detector.

anomalyDetectorSummary_anomalyDetectorName :: Lens' AnomalyDetectorSummary (Maybe Text)

The name of the detector.

anomalyDetectorSummary_anomalyDetectorDescription :: Lens' AnomalyDetectorSummary (Maybe Text)

A description of the detector.

anomalyDetectorSummary_tags :: Lens' AnomalyDetectorSummary (Maybe (HashMap Text Text))

The detector's tags.

anomalyDetectorSummary_lastModificationTime :: Lens' AnomalyDetectorSummary (Maybe UTCTime)

The time at which the detector was last modified.

AnomalyGroup

anomalyGroup_metricLevelImpactList :: Lens' AnomalyGroup (Maybe [MetricLevelImpact])

A list of measures affected by the anomaly.

anomalyGroup_startTime :: Lens' AnomalyGroup (Maybe Text)

The start time for the group.

anomalyGroup_anomalyGroupId :: Lens' AnomalyGroup (Maybe Text)

The ID of the anomaly group.

anomalyGroup_anomalyGroupScore :: Lens' AnomalyGroup (Maybe Double)

The severity score of the group.

anomalyGroup_primaryMetricName :: Lens' AnomalyGroup (Maybe Text)

The name of the primary affected measure for the group.

anomalyGroup_endTime :: Lens' AnomalyGroup (Maybe Text)

The end time for the group.

AnomalyGroupStatistics

anomalyGroupStatistics_itemizedMetricStatsList :: Lens' AnomalyGroupStatistics (Maybe [ItemizedMetricStats])

Statistics for individual metrics within the group.

anomalyGroupStatistics_evaluationStartDate :: Lens' AnomalyGroupStatistics (Maybe Text)

The start of the time range that was searched.

anomalyGroupStatistics_totalCount :: Lens' AnomalyGroupStatistics (Maybe Int)

The number of groups found.

AnomalyGroupSummary

anomalyGroupSummary_startTime :: Lens' AnomalyGroupSummary (Maybe Text)

The start time for the group.

anomalyGroupSummary_anomalyGroupId :: Lens' AnomalyGroupSummary (Maybe Text)

The ID of the anomaly group.

anomalyGroupSummary_anomalyGroupScore :: Lens' AnomalyGroupSummary (Maybe Double)

The severity score of the group.

anomalyGroupSummary_primaryMetricName :: Lens' AnomalyGroupSummary (Maybe Text)

The name of the primary affected measure for the group.

anomalyGroupSummary_endTime :: Lens' AnomalyGroupSummary (Maybe Text)

The end time for the group.

AnomalyGroupTimeSeries

anomalyGroupTimeSeries_timeSeriesId :: Lens' AnomalyGroupTimeSeries (Maybe Text)

The ID of the metric.

anomalyGroupTimeSeries_anomalyGroupId :: Lens' AnomalyGroupTimeSeries Text

The ID of the anomaly group.

AnomalyGroupTimeSeriesFeedback

anomalyGroupTimeSeriesFeedback_anomalyGroupId :: Lens' AnomalyGroupTimeSeriesFeedback Text

The ID of the anomaly group.

anomalyGroupTimeSeriesFeedback_timeSeriesId :: Lens' AnomalyGroupTimeSeriesFeedback Text

The ID of the metric.

anomalyGroupTimeSeriesFeedback_isAnomaly :: Lens' AnomalyGroupTimeSeriesFeedback Bool

Feedback on whether the metric is a legitimate anomaly.

AppFlowConfig

appFlowConfig_roleArn :: Lens' AppFlowConfig Text

An IAM role that gives Amazon Lookout for Metrics permission to access the flow.

appFlowConfig_flowName :: Lens' AppFlowConfig Text

name of the flow.

CloudWatchConfig

cloudWatchConfig_roleArn :: Lens' CloudWatchConfig Text

An IAM role that gives Amazon Lookout for Metrics permission to access data in Amazon CloudWatch.

ContributionMatrix

contributionMatrix_dimensionContributionList :: Lens' ContributionMatrix (Maybe [DimensionContribution])

A list of contributing dimensions.

CsvFormatDescriptor

csvFormatDescriptor_quoteSymbol :: Lens' CsvFormatDescriptor (Maybe Text)

The character used as a quote character.

csvFormatDescriptor_containsHeader :: Lens' CsvFormatDescriptor (Maybe Bool)

Whether or not the source CSV file contains a header.

csvFormatDescriptor_charset :: Lens' CsvFormatDescriptor (Maybe Text)

The character set in which the source CSV file is written.

csvFormatDescriptor_headerList :: Lens' CsvFormatDescriptor (Maybe [Text])

A list of the source CSV file's headers, if any.

csvFormatDescriptor_fileCompression :: Lens' CsvFormatDescriptor (Maybe CSVFileCompression)

The level of compression of the source CSV file.

csvFormatDescriptor_delimiter :: Lens' CsvFormatDescriptor (Maybe Text)

The character used to delimit the source CSV file.

DimensionContribution

dimensionContribution_dimensionValueContributionList :: Lens' DimensionContribution (Maybe [DimensionValueContribution])

A list of dimension values that contributed to the anomaly.

dimensionContribution_dimensionName :: Lens' DimensionContribution (Maybe Text)

The name of the dimension.

DimensionNameValue

dimensionNameValue_dimensionName :: Lens' DimensionNameValue Text

The name of the dimension.

dimensionNameValue_dimensionValue :: Lens' DimensionNameValue Text

The value of the dimension.

DimensionValueContribution

dimensionValueContribution_dimensionValue :: Lens' DimensionValueContribution (Maybe Text)

The value of the dimension.

dimensionValueContribution_contributionScore :: Lens' DimensionValueContribution (Maybe Double)

The severity score of the value.

ExecutionStatus

executionStatus_status :: Lens' ExecutionStatus (Maybe AnomalyDetectionTaskStatus)

The run's status.

executionStatus_failureReason :: Lens' ExecutionStatus (Maybe Text)

The reason that the run failed, if applicable.

executionStatus_timestamp :: Lens' ExecutionStatus (Maybe Text)

The run's timestamp.

FileFormatDescriptor

fileFormatDescriptor_jsonFormatDescriptor :: Lens' FileFormatDescriptor (Maybe JsonFormatDescriptor)

Contains information about how a source JSON data file should be analyzed.

fileFormatDescriptor_csvFormatDescriptor :: Lens' FileFormatDescriptor (Maybe CsvFormatDescriptor)

Contains information about how a source CSV data file should be analyzed.

ItemizedMetricStats

itemizedMetricStats_metricName :: Lens' ItemizedMetricStats (Maybe Text)

The name of the measure.

itemizedMetricStats_occurrenceCount :: Lens' ItemizedMetricStats (Maybe Int)

The number of times that the measure appears.

JsonFormatDescriptor

jsonFormatDescriptor_charset :: Lens' JsonFormatDescriptor (Maybe Text)

The character set in which the source JSON file is written.

jsonFormatDescriptor_fileCompression :: Lens' JsonFormatDescriptor (Maybe JsonFileCompression)

The level of compression of the source CSV file.

LambdaConfiguration

lambdaConfiguration_roleArn :: Lens' LambdaConfiguration Text

The ARN of an IAM role that has permission to invoke the Lambda function.

lambdaConfiguration_lambdaArn :: Lens' LambdaConfiguration Text

The ARN of the Lambda function.

Metric

metric_namespace :: Lens' Metric (Maybe Text)

The namespace for the metric.

metric_metricName :: Lens' Metric Text

The name of the metric.

metric_aggregationFunction :: Lens' Metric AggregationFunction

The function with which the metric is calculated.

MetricLevelImpact

metricLevelImpact_contributionMatrix :: Lens' MetricLevelImpact (Maybe ContributionMatrix)

Details about the dimensions that contributed to the anomaly.

metricLevelImpact_metricName :: Lens' MetricLevelImpact (Maybe Text)

The name of the measure.

metricLevelImpact_numTimeSeries :: Lens' MetricLevelImpact (Maybe Int)

The number of anomalous metrics for the measure.

MetricSetSummary

metricSetSummary_creationTime :: Lens' MetricSetSummary (Maybe UTCTime)

The time at which the dataset was created.

metricSetSummary_anomalyDetectorArn :: Lens' MetricSetSummary (Maybe Text)

The ARN of the detector to which the dataset belongs.

metricSetSummary_metricSetName :: Lens' MetricSetSummary (Maybe Text)

The name of the dataset.

metricSetSummary_metricSetDescription :: Lens' MetricSetSummary (Maybe Text)

The description of the dataset.

metricSetSummary_metricSetArn :: Lens' MetricSetSummary (Maybe Text)

The ARN of the dataset.

metricSetSummary_tags :: Lens' MetricSetSummary (Maybe (HashMap Text Text))

The dataset's tags.

metricSetSummary_lastModificationTime :: Lens' MetricSetSummary (Maybe UTCTime)

The time at which the dataset was last modified.

MetricSource

metricSource_redshiftSourceConfig :: Lens' MetricSource (Maybe RedshiftSourceConfig)

An object containing information about the Amazon Redshift database configuration.

metricSource_s3SourceConfig :: Lens' MetricSource (Maybe S3SourceConfig)

Undocumented member.

metricSource_rDSSourceConfig :: Lens' MetricSource (Maybe RDSSourceConfig)

An object containing information about the Amazon Relational Database Service (RDS) configuration.

metricSource_appFlowConfig :: Lens' MetricSource (Maybe AppFlowConfig)

An object containing information about the AppFlow configuration.

metricSource_cloudWatchConfig :: Lens' MetricSource (Maybe CloudWatchConfig)

An object containing information about the Amazon CloudWatch monitoring configuration.

RDSSourceConfig

rDSSourceConfig_dbInstanceIdentifier :: Lens' RDSSourceConfig Text

A string identifying the database instance.

rDSSourceConfig_databaseHost :: Lens' RDSSourceConfig Text

The host name of the database.

rDSSourceConfig_databasePort :: Lens' RDSSourceConfig Natural

The port number where the database can be accessed.

rDSSourceConfig_secretManagerArn :: Lens' RDSSourceConfig Text

The Amazon Resource Name (ARN) of the AWS Secrets Manager role.

rDSSourceConfig_databaseName :: Lens' RDSSourceConfig Text

The name of the RDS database.

rDSSourceConfig_tableName :: Lens' RDSSourceConfig Text

The name of the table in the database.

rDSSourceConfig_roleArn :: Lens' RDSSourceConfig Text

The Amazon Resource Name (ARN) of the role.

rDSSourceConfig_vpcConfiguration :: Lens' RDSSourceConfig VpcConfiguration

An object containing information about the Amazon Virtual Private Cloud (VPC) configuration.

RedshiftSourceConfig

redshiftSourceConfig_clusterIdentifier :: Lens' RedshiftSourceConfig Text

A string identifying the Redshift cluster.

redshiftSourceConfig_databaseHost :: Lens' RedshiftSourceConfig Text

The name of the database host.

redshiftSourceConfig_databasePort :: Lens' RedshiftSourceConfig Natural

The port number where the database can be accessed.

redshiftSourceConfig_secretManagerArn :: Lens' RedshiftSourceConfig Text

The Amazon Resource Name (ARN) of the AWS Secrets Manager role.

redshiftSourceConfig_databaseName :: Lens' RedshiftSourceConfig Text

The Redshift database name.

redshiftSourceConfig_tableName :: Lens' RedshiftSourceConfig Text

The table name of the Redshift database.

redshiftSourceConfig_roleArn :: Lens' RedshiftSourceConfig Text

The Amazon Resource Name (ARN) of the role providing access to the database.

redshiftSourceConfig_vpcConfiguration :: Lens' RedshiftSourceConfig VpcConfiguration

Contains information about the Amazon Virtual Private Cloud (VPC) configuration.

S3SourceConfig

s3SourceConfig_templatedPathList :: Lens' S3SourceConfig (Maybe (NonEmpty Text))

A list of templated paths to the source files.

s3SourceConfig_historicalDataPathList :: Lens' S3SourceConfig (Maybe (NonEmpty Text))

A list of paths to the historical data files.

s3SourceConfig_fileFormatDescriptor :: Lens' S3SourceConfig (Maybe FileFormatDescriptor)

Contains information about a source file's formatting.

s3SourceConfig_roleArn :: Lens' S3SourceConfig Text

The ARN of an IAM role that has read and write access permissions to the source S3 bucket.

SNSConfiguration

sNSConfiguration_roleArn :: Lens' SNSConfiguration Text

The ARN of the IAM role that has access to the target SNS topic.

sNSConfiguration_snsTopicArn :: Lens' SNSConfiguration Text

The ARN of the target SNS topic.

SampleDataS3SourceConfig

sampleDataS3SourceConfig_templatedPathList :: Lens' SampleDataS3SourceConfig (Maybe (NonEmpty Text))

An array of strings containing the list of templated paths.

sampleDataS3SourceConfig_historicalDataPathList :: Lens' SampleDataS3SourceConfig (Maybe (NonEmpty Text))

An array of strings containing the historical set of data paths.

sampleDataS3SourceConfig_roleArn :: Lens' SampleDataS3SourceConfig Text

The Amazon Resource Name (ARN) of the role.

sampleDataS3SourceConfig_fileFormatDescriptor :: Lens' SampleDataS3SourceConfig FileFormatDescriptor

Undocumented member.

TimeSeries

timeSeries_timeSeriesId :: Lens' TimeSeries Text

The ID of the metric.

timeSeries_dimensionList :: Lens' TimeSeries [DimensionNameValue]

The dimensions of the metric.

timeSeries_metricValueList :: Lens' TimeSeries [Double]

The values for the metric.

TimeSeriesFeedback

timeSeriesFeedback_isAnomaly :: Lens' TimeSeriesFeedback (Maybe Bool)

Feedback on whether the metric is a legitimate anomaly.

timeSeriesFeedback_timeSeriesId :: Lens' TimeSeriesFeedback (Maybe Text)

The ID of the metric.

TimestampColumn

timestampColumn_columnFormat :: Lens' TimestampColumn (Maybe Text)

The format of the timestamp column.

timestampColumn_columnName :: Lens' TimestampColumn (Maybe Text)

The name of the timestamp column.

VpcConfiguration

vpcConfiguration_subnetIdList :: Lens' VpcConfiguration [Text]

An array of strings containing the Amazon VPC subnet IDs (e.g., subnet-0bb1c79de3EXAMPLE.

vpcConfiguration_securityGroupIdList :: Lens' VpcConfiguration [Text]

An array of strings containing the list of security groups.