Network.AWS.LookoutMetrics.Types.AnomalyDetectorSummary

Description

Documentation

data AnomalyDetectorSummary

Contains information about an an anomaly detector.

See: newAnomalyDetectorSummary smart constructor.

Constructors

AnomalyDetectorSummary'

Fields

creationTime :: Maybe POSIX

The time at which the detector was created.

status :: Maybe AnomalyDetectorStatus

The status of detector.

anomalyDetectorArn :: Maybe Text

The ARN of the detector.

anomalyDetectorName :: Maybe Text

The name of the detector.

anomalyDetectorDescription :: Maybe Text

A description of the detector.

tags :: Maybe (HashMap Text Text)

The detector's tags.

lastModificationTime :: Maybe POSIX

The time at which the detector was last modified.

newAnomalyDetectorSummary :: AnomalyDetectorSummary

Create a value of AnomalyDetectorSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:AnomalyDetectorSummary', anomalyDetectorSummary_creationTime - The time at which the detector was created.

$sel:status:AnomalyDetectorSummary', anomalyDetectorSummary_status - The status of detector.

$sel:anomalyDetectorArn:AnomalyDetectorSummary', anomalyDetectorSummary_anomalyDetectorArn - The ARN of the detector.

$sel:anomalyDetectorName:AnomalyDetectorSummary', anomalyDetectorSummary_anomalyDetectorName - The name of the detector.

$sel:anomalyDetectorDescription:AnomalyDetectorSummary', anomalyDetectorSummary_anomalyDetectorDescription - A description of the detector.

$sel:tags:AnomalyDetectorSummary', anomalyDetectorSummary_tags - The detector's tags.

$sel:lastModificationTime:AnomalyDetectorSummary', anomalyDetectorSummary_lastModificationTime - The time at which the detector was last modified.

anomalyDetectorSummary_creationTime :: Lens' AnomalyDetectorSummary (Maybe UTCTime)

The time at which the detector was created.

anomalyDetectorSummary_status :: Lens' AnomalyDetectorSummary (Maybe AnomalyDetectorStatus)

The status of detector.

anomalyDetectorSummary_anomalyDetectorArn :: Lens' AnomalyDetectorSummary (Maybe Text)

The ARN of the detector.

anomalyDetectorSummary_anomalyDetectorName :: Lens' AnomalyDetectorSummary (Maybe Text)

The name of the detector.

anomalyDetectorSummary_anomalyDetectorDescription :: Lens' AnomalyDetectorSummary (Maybe Text)

A description of the detector.

anomalyDetectorSummary_tags :: Lens' AnomalyDetectorSummary (Maybe (HashMap Text Text))

The detector's tags.

anomalyDetectorSummary_lastModificationTime :: Lens' AnomalyDetectorSummary (Maybe UTCTime)

The time at which the detector was last modified.