Network.AWS.LookoutMetrics.Types.AnomalyGroup

Description

Documentation

data AnomalyGroup

A group of anomalous metrics

See: newAnomalyGroup smart constructor.

Constructors

AnomalyGroup'

Fields

metricLevelImpactList :: Maybe [MetricLevelImpact]

A list of measures affected by the anomaly.

startTime :: Maybe Text

The start time for the group.

anomalyGroupId :: Maybe Text

The ID of the anomaly group.

anomalyGroupScore :: Maybe Double

The severity score of the group.

primaryMetricName :: Maybe Text

The name of the primary affected measure for the group.

endTime :: Maybe Text

The end time for the group.

newAnomalyGroup :: AnomalyGroup

Create a value of AnomalyGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricLevelImpactList:AnomalyGroup', anomalyGroup_metricLevelImpactList - A list of measures affected by the anomaly.

$sel:startTime:AnomalyGroup', anomalyGroup_startTime - The start time for the group.

$sel:anomalyGroupId:AnomalyGroup', anomalyGroup_anomalyGroupId - The ID of the anomaly group.

$sel:anomalyGroupScore:AnomalyGroup', anomalyGroup_anomalyGroupScore - The severity score of the group.

$sel:primaryMetricName:AnomalyGroup', anomalyGroup_primaryMetricName - The name of the primary affected measure for the group.

$sel:endTime:AnomalyGroup', anomalyGroup_endTime - The end time for the group.

anomalyGroup_metricLevelImpactList :: Lens' AnomalyGroup (Maybe [MetricLevelImpact])

A list of measures affected by the anomaly.

anomalyGroup_startTime :: Lens' AnomalyGroup (Maybe Text)

The start time for the group.

anomalyGroup_anomalyGroupId :: Lens' AnomalyGroup (Maybe Text)

The ID of the anomaly group.

anomalyGroup_anomalyGroupScore :: Lens' AnomalyGroup (Maybe Double)

The severity score of the group.

anomalyGroup_primaryMetricName :: Lens' AnomalyGroup (Maybe Text)

The name of the primary affected measure for the group.

anomalyGroup_endTime :: Lens' AnomalyGroup (Maybe Text)

The end time for the group.