Network.AWS.LookoutMetrics.Types.MetricSource

Description

Documentation

data MetricSource

Contains information about source data used to generate a metric.

See: newMetricSource smart constructor.

Constructors

MetricSource'

Fields

redshiftSourceConfig :: Maybe RedshiftSourceConfig

An object containing information about the Amazon Redshift database configuration.

s3SourceConfig :: Maybe S3SourceConfig

rDSSourceConfig :: Maybe RDSSourceConfig

An object containing information about the Amazon Relational Database Service (RDS) configuration.

appFlowConfig :: Maybe AppFlowConfig

An object containing information about the AppFlow configuration.

cloudWatchConfig :: Maybe CloudWatchConfig

An object containing information about the Amazon CloudWatch monitoring configuration.

newMetricSource :: MetricSource

Create a value of MetricSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:redshiftSourceConfig:MetricSource', metricSource_redshiftSourceConfig - An object containing information about the Amazon Redshift database configuration.

$sel:s3SourceConfig:MetricSource', metricSource_s3SourceConfig - Undocumented member.

$sel:rDSSourceConfig:MetricSource', metricSource_rDSSourceConfig - An object containing information about the Amazon Relational Database Service (RDS) configuration.

$sel:appFlowConfig:MetricSource', metricSource_appFlowConfig - An object containing information about the AppFlow configuration.

$sel:cloudWatchConfig:MetricSource', metricSource_cloudWatchConfig - An object containing information about the Amazon CloudWatch monitoring configuration.

metricSource_redshiftSourceConfig :: Lens' MetricSource (Maybe RedshiftSourceConfig)

An object containing information about the Amazon Redshift database configuration.

metricSource_s3SourceConfig :: Lens' MetricSource (Maybe S3SourceConfig)

Undocumented member.

metricSource_rDSSourceConfig :: Lens' MetricSource (Maybe RDSSourceConfig)

An object containing information about the Amazon Relational Database Service (RDS) configuration.

metricSource_appFlowConfig :: Lens' MetricSource (Maybe AppFlowConfig)

An object containing information about the AppFlow configuration.

metricSource_cloudWatchConfig :: Lens' MetricSource (Maybe CloudWatchConfig)

An object containing information about the Amazon CloudWatch monitoring configuration.