Network.AWS.LookoutMetrics.Types.RDSSourceConfig

Description

Documentation

data RDSSourceConfig

Contains information about the Amazon Relational Database Service (RDS) configuration.

See: newRDSSourceConfig smart constructor.

Constructors

RDSSourceConfig'

Fields

dbInstanceIdentifier :: Text

A string identifying the database instance.

databaseHost :: Text

The host name of the database.

databasePort :: Natural

The port number where the database can be accessed.

secretManagerArn :: Text

The Amazon Resource Name (ARN) of the AWS Secrets Manager role.

databaseName :: Text

The name of the RDS database.

tableName :: Text

The name of the table in the database.

roleArn :: Text

The Amazon Resource Name (ARN) of the role.

vpcConfiguration :: VpcConfiguration

An object containing information about the Amazon Virtual Private Cloud (VPC) configuration.

newRDSSourceConfig

Arguments

:: Text
$sel:dbInstanceIdentifier:RDSSourceConfig'

-> Text
$sel:databaseHost:RDSSourceConfig'

-> Natural
$sel:databasePort:RDSSourceConfig'

-> Text
$sel:secretManagerArn:RDSSourceConfig'

-> Text
$sel:databaseName:RDSSourceConfig'

-> Text
$sel:tableName:RDSSourceConfig'

-> Text
$sel:roleArn:RDSSourceConfig'

-> VpcConfiguration
$sel:vpcConfiguration:RDSSourceConfig'

-> RDSSourceConfig

Create a value of RDSSourceConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbInstanceIdentifier:RDSSourceConfig', rDSSourceConfig_dbInstanceIdentifier - A string identifying the database instance.

$sel:databaseHost:RDSSourceConfig', rDSSourceConfig_databaseHost - The host name of the database.

$sel:databasePort:RDSSourceConfig', rDSSourceConfig_databasePort - The port number where the database can be accessed.

$sel:secretManagerArn:RDSSourceConfig', rDSSourceConfig_secretManagerArn - The Amazon Resource Name (ARN) of the AWS Secrets Manager role.

$sel:databaseName:RDSSourceConfig', rDSSourceConfig_databaseName - The name of the RDS database.

$sel:tableName:RDSSourceConfig', rDSSourceConfig_tableName - The name of the table in the database.

$sel:roleArn:RDSSourceConfig', rDSSourceConfig_roleArn - The Amazon Resource Name (ARN) of the role.

$sel:vpcConfiguration:RDSSourceConfig', rDSSourceConfig_vpcConfiguration - An object containing information about the Amazon Virtual Private Cloud (VPC) configuration.

rDSSourceConfig_dbInstanceIdentifier :: Lens' RDSSourceConfig Text

A string identifying the database instance.

rDSSourceConfig_databaseHost :: Lens' RDSSourceConfig Text

The host name of the database.

rDSSourceConfig_databasePort :: Lens' RDSSourceConfig Natural

The port number where the database can be accessed.

rDSSourceConfig_secretManagerArn :: Lens' RDSSourceConfig Text

The Amazon Resource Name (ARN) of the AWS Secrets Manager role.

rDSSourceConfig_databaseName :: Lens' RDSSourceConfig Text

The name of the RDS database.

rDSSourceConfig_tableName :: Lens' RDSSourceConfig Text

The name of the table in the database.

rDSSourceConfig_roleArn :: Lens' RDSSourceConfig Text

The Amazon Resource Name (ARN) of the role.

rDSSourceConfig_vpcConfiguration :: Lens' RDSSourceConfig VpcConfiguration

An object containing information about the Amazon Virtual Private Cloud (VPC) configuration.