Network.AWS.LookoutMetrics.UpdateAnomalyDetector

Description

Updates a detector. After activation, you can only change a detector's ingestion delay and description.

Creating a Request

data UpdateAnomalyDetector

See: newUpdateAnomalyDetector smart constructor.

Constructors

UpdateAnomalyDetector'

Fields

kmsKeyArn :: Maybe Text

The Amazon Resource Name (ARN) of an AWS KMS encryption key.

anomalyDetectorConfig :: Maybe AnomalyDetectorConfig

Contains information about the configuration to which the detector will be updated.

anomalyDetectorDescription :: Maybe Text

The updated detector description.

anomalyDetectorArn :: Text

The ARN of the detector to update.

newUpdateAnomalyDetector

Arguments

:: Text
$sel:anomalyDetectorArn:UpdateAnomalyDetector'

-> UpdateAnomalyDetector

Create a value of UpdateAnomalyDetector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyArn:UpdateAnomalyDetector', updateAnomalyDetector_kmsKeyArn - The Amazon Resource Name (ARN) of an AWS KMS encryption key.

$sel:anomalyDetectorConfig:UpdateAnomalyDetector', updateAnomalyDetector_anomalyDetectorConfig - Contains information about the configuration to which the detector will be updated.

$sel:anomalyDetectorDescription:UpdateAnomalyDetector', updateAnomalyDetector_anomalyDetectorDescription - The updated detector description.

$sel:anomalyDetectorArn:UpdateAnomalyDetector', updateAnomalyDetector_anomalyDetectorArn - The ARN of the detector to update.

Request Lenses

updateAnomalyDetector_kmsKeyArn :: Lens' UpdateAnomalyDetector (Maybe Text)

The Amazon Resource Name (ARN) of an AWS KMS encryption key.

updateAnomalyDetector_anomalyDetectorConfig :: Lens' UpdateAnomalyDetector (Maybe AnomalyDetectorConfig)

Contains information about the configuration to which the detector will be updated.

updateAnomalyDetector_anomalyDetectorDescription :: Lens' UpdateAnomalyDetector (Maybe Text)

The updated detector description.

updateAnomalyDetector_anomalyDetectorArn :: Lens' UpdateAnomalyDetector Text

The ARN of the detector to update.

Destructuring the Response

data UpdateAnomalyDetectorResponse

See: newUpdateAnomalyDetectorResponse smart constructor.

Constructors

UpdateAnomalyDetectorResponse'

Fields

anomalyDetectorArn :: Maybe Text

The ARN of the updated detector.

httpStatus :: Int

The response's http status code.

newUpdateAnomalyDetectorResponse

Arguments

:: Int
$sel:httpStatus:UpdateAnomalyDetectorResponse'

-> UpdateAnomalyDetectorResponse

Create a value of UpdateAnomalyDetectorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:anomalyDetectorArn:UpdateAnomalyDetector', updateAnomalyDetectorResponse_anomalyDetectorArn - The ARN of the updated detector.

$sel:httpStatus:UpdateAnomalyDetectorResponse', updateAnomalyDetectorResponse_httpStatus - The response's http status code.

Response Lenses

updateAnomalyDetectorResponse_anomalyDetectorArn :: Lens' UpdateAnomalyDetectorResponse (Maybe Text)

The ARN of the updated detector.

updateAnomalyDetectorResponse_httpStatus :: Lens' UpdateAnomalyDetectorResponse Int

The response's http status code.