Network.AWS.LookoutVision.CreateDataset

Description

Creates a new dataset in an Amazon Lookout for Vision project. CreateDataset can create a training or a test dataset from a valid dataset source (DatasetSource).

If you want a single dataset project, specify train for the value of DatasetType.

To have a project with separate training and test datasets, call CreateDataset twice. On the first call, specify train for the value of DatasetType. On the second call, specify test for the value of DatasetType.

This operation requires permissions to perform the lookoutvision:CreateDataset operation.

Creating a Request

data CreateDataset

See: newCreateDataset smart constructor.

Constructors

CreateDataset'

Fields

clientToken :: Maybe Text

ClientToken is an idempotency token that ensures a call to CreateDataset completes only once. You choose the value to pass. For example, An issue, such as an network outage, might prevent you from getting a response from CreateDataset. In this case, safely retry your call to CreateDataset by using the same ClientToken parameter value. An error occurs if the other input parameters are not the same as in the first request. Using a different value for ClientToken is considered a new call to CreateDataset. An idempotency token is active for 8 hours.

datasetSource :: Maybe DatasetSource

The location of the manifest file that Amazon Lookout for Vision uses to create the dataset.

If you don't specify DatasetSource, an empty dataset is created and the operation synchronously returns. Later, you can add JSON Lines by calling UpdateDatasetEntries.

If you specify a value for DataSource, the manifest at the S3 location is validated and used to create the dataset. The call to CreateDataset is asynchronous and might take a while to complete. To find out the current status, Check the value of Status returned in a call to DescribeDataset.

projectName :: Text

The name of the project in which you want to create a dataset.

datasetType :: Text

The type of the dataset. Specify train for a training dataset. Specify test for a test dataset.

newCreateDataset

Arguments

:: Text
$sel:projectName:CreateDataset'

-> Text
$sel:datasetType:CreateDataset'

-> CreateDataset

Create a value of CreateDataset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:CreateDataset', createDataset_clientToken - ClientToken is an idempotency token that ensures a call to CreateDataset completes only once. You choose the value to pass. For example, An issue, such as an network outage, might prevent you from getting a response from CreateDataset. In this case, safely retry your call to CreateDataset by using the same ClientToken parameter value. An error occurs if the other input parameters are not the same as in the first request. Using a different value for ClientToken is considered a new call to CreateDataset. An idempotency token is active for 8 hours.

$sel:datasetSource:CreateDataset', createDataset_datasetSource - The location of the manifest file that Amazon Lookout for Vision uses to create the dataset.

If you don't specify DatasetSource, an empty dataset is created and the operation synchronously returns. Later, you can add JSON Lines by calling UpdateDatasetEntries.

If you specify a value for DataSource, the manifest at the S3 location is validated and used to create the dataset. The call to CreateDataset is asynchronous and might take a while to complete. To find out the current status, Check the value of Status returned in a call to DescribeDataset.

$sel:projectName:CreateDataset', createDataset_projectName - The name of the project in which you want to create a dataset.

$sel:datasetType:CreateDataset', createDataset_datasetType - The type of the dataset. Specify train for a training dataset. Specify test for a test dataset.

Request Lenses

createDataset_clientToken :: Lens' CreateDataset (Maybe Text)

ClientToken is an idempotency token that ensures a call to CreateDataset completes only once. You choose the value to pass. For example, An issue, such as an network outage, might prevent you from getting a response from CreateDataset. In this case, safely retry your call to CreateDataset by using the same ClientToken parameter value. An error occurs if the other input parameters are not the same as in the first request. Using a different value for ClientToken is considered a new call to CreateDataset. An idempotency token is active for 8 hours.

createDataset_datasetSource :: Lens' CreateDataset (Maybe DatasetSource)

The location of the manifest file that Amazon Lookout for Vision uses to create the dataset.

If you don't specify DatasetSource, an empty dataset is created and the operation synchronously returns. Later, you can add JSON Lines by calling UpdateDatasetEntries.

If you specify a value for DataSource, the manifest at the S3 location is validated and used to create the dataset. The call to CreateDataset is asynchronous and might take a while to complete. To find out the current status, Check the value of Status returned in a call to DescribeDataset.

createDataset_projectName :: Lens' CreateDataset Text

The name of the project in which you want to create a dataset.

createDataset_datasetType :: Lens' CreateDataset Text

The type of the dataset. Specify train for a training dataset. Specify test for a test dataset.

Destructuring the Response

data CreateDatasetResponse

See: newCreateDatasetResponse smart constructor.

Constructors

CreateDatasetResponse'

Fields

datasetMetadata :: Maybe DatasetMetadata

Information about the dataset.

httpStatus :: Int

The response's http status code.

newCreateDatasetResponse

Arguments

:: Int
$sel:httpStatus:CreateDatasetResponse'

-> CreateDatasetResponse

Create a value of CreateDatasetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetMetadata:CreateDatasetResponse', createDatasetResponse_datasetMetadata - Information about the dataset.

$sel:httpStatus:CreateDatasetResponse', createDatasetResponse_httpStatus - The response's http status code.

Response Lenses

createDatasetResponse_datasetMetadata :: Lens' CreateDatasetResponse (Maybe DatasetMetadata)

Information about the dataset.

createDatasetResponse_httpStatus :: Lens' CreateDatasetResponse Int

The response's http status code.