Network.AWS.LookoutVision.ListDatasetEntries

Description

Lists the JSON Lines within a dataset. An Amazon Lookout for Vision JSON Line contains the anomaly information for a single image, including the image location and the assigned label.

This operation requires permissions to perform the lookoutvision:ListDatasetEntries operation.

This operation returns paginated results.

Creating a Request

data ListDatasetEntries

See: newListDatasetEntries smart constructor.

Constructors

ListDatasetEntries'

Fields

beforeCreationDate :: Maybe POSIX

Only includes entries before the specified date in the response. For example, 2020-06-23T00:00:00.

sourceRefContains :: Maybe Text

Perform a "contains" search on the values of the source-ref key within the dataset. For example a value of "IMG_17" returns all JSON Lines where the source-ref key value matches *IMG_17*.

nextToken :: Maybe Text

If the previous response was incomplete (because there is more data to retrieve), Amazon Lookout for Vision returns a pagination token in the response. You can use this pagination token to retrieve the next set of dataset entries.

labeled :: Maybe Bool

Specify true to include labeled entries, otherwise specify false. If you don't specify a value, Lookout for Vision returns all entries.

anomalyClass :: Maybe Text

Specify normal to include only normal images. Specify anomaly to only include anomalous entries. If you don't specify a value, Amazon Lookout for Vision returns normal and anomalous images.

maxResults :: Maybe Natural

The maximum number of results to return per paginated call. The largest value you can specify is 100. If you specify a value greater than 100, a ValidationException error occurs. The default value is 100.

afterCreationDate :: Maybe POSIX

Only includes entries after the specified date in the response. For example, 2020-06-23T00:00:00.

projectName :: Text

The name of the project that contains the dataset that you want to list.

datasetType :: Text

The type of the dataset that you want to list. Specify train to list the training dataset. Specify test to list the test dataset. If you have a single dataset project, specify train.

newListDatasetEntries

Arguments

:: Text
$sel:projectName:ListDatasetEntries'

-> Text
$sel:datasetType:ListDatasetEntries'

-> ListDatasetEntries

Create a value of ListDatasetEntries with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:beforeCreationDate:ListDatasetEntries', listDatasetEntries_beforeCreationDate - Only includes entries before the specified date in the response. For example, 2020-06-23T00:00:00.

$sel:sourceRefContains:ListDatasetEntries', listDatasetEntries_sourceRefContains - Perform a "contains" search on the values of the source-ref key within the dataset. For example a value of "IMG_17" returns all JSON Lines where the source-ref key value matches *IMG_17*.

$sel:nextToken:ListDatasetEntries', listDatasetEntries_nextToken - If the previous response was incomplete (because there is more data to retrieve), Amazon Lookout for Vision returns a pagination token in the response. You can use this pagination token to retrieve the next set of dataset entries.

$sel:labeled:ListDatasetEntries', listDatasetEntries_labeled - Specify true to include labeled entries, otherwise specify false. If you don't specify a value, Lookout for Vision returns all entries.

$sel:anomalyClass:ListDatasetEntries', listDatasetEntries_anomalyClass - Specify normal to include only normal images. Specify anomaly to only include anomalous entries. If you don't specify a value, Amazon Lookout for Vision returns normal and anomalous images.

$sel:maxResults:ListDatasetEntries', listDatasetEntries_maxResults - The maximum number of results to return per paginated call. The largest value you can specify is 100. If you specify a value greater than 100, a ValidationException error occurs. The default value is 100.

$sel:afterCreationDate:ListDatasetEntries', listDatasetEntries_afterCreationDate - Only includes entries after the specified date in the response. For example, 2020-06-23T00:00:00.

$sel:projectName:ListDatasetEntries', listDatasetEntries_projectName - The name of the project that contains the dataset that you want to list.

$sel:datasetType:ListDatasetEntries', listDatasetEntries_datasetType - The type of the dataset that you want to list. Specify train to list the training dataset. Specify test to list the test dataset. If you have a single dataset project, specify train.

Request Lenses

listDatasetEntries_beforeCreationDate :: Lens' ListDatasetEntries (Maybe UTCTime)

Only includes entries before the specified date in the response. For example, 2020-06-23T00:00:00.

listDatasetEntries_sourceRefContains :: Lens' ListDatasetEntries (Maybe Text)

Perform a "contains" search on the values of the source-ref key within the dataset. For example a value of "IMG_17" returns all JSON Lines where the source-ref key value matches *IMG_17*.

listDatasetEntries_nextToken :: Lens' ListDatasetEntries (Maybe Text)

If the previous response was incomplete (because there is more data to retrieve), Amazon Lookout for Vision returns a pagination token in the response. You can use this pagination token to retrieve the next set of dataset entries.

listDatasetEntries_labeled :: Lens' ListDatasetEntries (Maybe Bool)

Specify true to include labeled entries, otherwise specify false. If you don't specify a value, Lookout for Vision returns all entries.

listDatasetEntries_anomalyClass :: Lens' ListDatasetEntries (Maybe Text)

Specify normal to include only normal images. Specify anomaly to only include anomalous entries. If you don't specify a value, Amazon Lookout for Vision returns normal and anomalous images.

listDatasetEntries_maxResults :: Lens' ListDatasetEntries (Maybe Natural)

The maximum number of results to return per paginated call. The largest value you can specify is 100. If you specify a value greater than 100, a ValidationException error occurs. The default value is 100.

listDatasetEntries_afterCreationDate :: Lens' ListDatasetEntries (Maybe UTCTime)

Only includes entries after the specified date in the response. For example, 2020-06-23T00:00:00.

listDatasetEntries_projectName :: Lens' ListDatasetEntries Text

The name of the project that contains the dataset that you want to list.

listDatasetEntries_datasetType :: Lens' ListDatasetEntries Text

The type of the dataset that you want to list. Specify train to list the training dataset. Specify test to list the test dataset. If you have a single dataset project, specify train.

Destructuring the Response

data ListDatasetEntriesResponse

See: newListDatasetEntriesResponse smart constructor.

Constructors

ListDatasetEntriesResponse'

Fields

datasetEntries :: Maybe [Text]

A list of the entries (JSON Lines) within the dataset.

nextToken :: Maybe Text

If the response is truncated, Amazon Lookout for Vision returns this token that you can use in the subsequent request to retrieve the next set ofdataset entries.

httpStatus :: Int

The response's http status code.

newListDatasetEntriesResponse

Arguments

:: Int
$sel:httpStatus:ListDatasetEntriesResponse'

-> ListDatasetEntriesResponse

Create a value of ListDatasetEntriesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetEntries:ListDatasetEntriesResponse', listDatasetEntriesResponse_datasetEntries - A list of the entries (JSON Lines) within the dataset.

$sel:nextToken:ListDatasetEntries', listDatasetEntriesResponse_nextToken - If the response is truncated, Amazon Lookout for Vision returns this token that you can use in the subsequent request to retrieve the next set ofdataset entries.

$sel:httpStatus:ListDatasetEntriesResponse', listDatasetEntriesResponse_httpStatus - The response's http status code.

Response Lenses

listDatasetEntriesResponse_datasetEntries :: Lens' ListDatasetEntriesResponse (Maybe [Text])

A list of the entries (JSON Lines) within the dataset.

listDatasetEntriesResponse_nextToken :: Lens' ListDatasetEntriesResponse (Maybe Text)

If the response is truncated, Amazon Lookout for Vision returns this token that you can use in the subsequent request to retrieve the next set ofdataset entries.

listDatasetEntriesResponse_httpStatus :: Lens' ListDatasetEntriesResponse Int

The response's http status code.