Network.AWS.LookoutVision.StartModel

Description

Starts the running of the version of an Amazon Lookout for Vision model. Starting a model takes a while to complete. To check the current state of the model, use DescribeModel.

A model is ready to use when its status is HOSTED.

Once the model is running, you can detect custom labels in new images by calling DetectAnomalies.

You are charged for the amount of time that the model is running. To stop a running model, call StopModel.

This operation requires permissions to perform the lookoutvision:StartModel operation.

Creating a Request

data StartModel

See: newStartModel smart constructor.

Constructors

StartModel'

Fields

clientToken :: Maybe Text

ClientToken is an idempotency token that ensures a call to StartModel completes only once. You choose the value to pass. For example, An issue, such as an network outage, might prevent you from getting a response from StartModel. In this case, safely retry your call to StartModel by using the same ClientToken parameter value. An error occurs if the other input parameters are not the same as in the first request. Using a different value for ClientToken is considered a new call to StartModel. An idempotency token is active for 8 hours.

projectName :: Text

The name of the project that contains the model that you want to start.

modelVersion :: Text

The version of the model that you want to start.

minInferenceUnits :: Natural

The minimum number of inference units to use. A single inference unit represents 1 hour of processing and can support up to 5 Transaction Pers Second (TPS). Use a higher number to increase the TPS throughput of your model. You are charged for the number of inference units that you use.

newStartModel

Arguments

:: Text
$sel:projectName:StartModel'

-> Text
$sel:modelVersion:StartModel'

-> Natural
$sel:minInferenceUnits:StartModel'

-> StartModel

Create a value of StartModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:StartModel', startModel_clientToken - ClientToken is an idempotency token that ensures a call to StartModel completes only once. You choose the value to pass. For example, An issue, such as an network outage, might prevent you from getting a response from StartModel. In this case, safely retry your call to StartModel by using the same ClientToken parameter value. An error occurs if the other input parameters are not the same as in the first request. Using a different value for ClientToken is considered a new call to StartModel. An idempotency token is active for 8 hours.

$sel:projectName:StartModel', startModel_projectName - The name of the project that contains the model that you want to start.

$sel:modelVersion:StartModel', startModel_modelVersion - The version of the model that you want to start.

$sel:minInferenceUnits:StartModel', startModel_minInferenceUnits - The minimum number of inference units to use. A single inference unit represents 1 hour of processing and can support up to 5 Transaction Pers Second (TPS). Use a higher number to increase the TPS throughput of your model. You are charged for the number of inference units that you use.

Request Lenses

startModel_clientToken :: Lens' StartModel (Maybe Text)

ClientToken is an idempotency token that ensures a call to StartModel completes only once. You choose the value to pass. For example, An issue, such as an network outage, might prevent you from getting a response from StartModel. In this case, safely retry your call to StartModel by using the same ClientToken parameter value. An error occurs if the other input parameters are not the same as in the first request. Using a different value for ClientToken is considered a new call to StartModel. An idempotency token is active for 8 hours.

startModel_projectName :: Lens' StartModel Text

The name of the project that contains the model that you want to start.

startModel_modelVersion :: Lens' StartModel Text

The version of the model that you want to start.

startModel_minInferenceUnits :: Lens' StartModel Natural

The minimum number of inference units to use. A single inference unit represents 1 hour of processing and can support up to 5 Transaction Pers Second (TPS). Use a higher number to increase the TPS throughput of your model. You are charged for the number of inference units that you use.

Destructuring the Response

data StartModelResponse

See: newStartModelResponse smart constructor.

Constructors

StartModelResponse'

Fields

status :: Maybe ModelHostingStatus

The current running status of the model.

httpStatus :: Int

The response's http status code.

newStartModelResponse

Arguments

:: Int
$sel:httpStatus:StartModelResponse'

-> StartModelResponse

Create a value of StartModelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:StartModelResponse', startModelResponse_status - The current running status of the model.

$sel:httpStatus:StartModelResponse', startModelResponse_httpStatus - The response's http status code.

Response Lenses

startModelResponse_status :: Lens' StartModelResponse (Maybe ModelHostingStatus)

The current running status of the model.

startModelResponse_httpStatus :: Lens' StartModelResponse Int

The response's http status code.