Network.AWS.LookoutVision.StopModel

Description

Stops the hosting of a running model. The operation might take a while to complete. To check the current status, call DescribeModel.

After the model hosting stops, the Status of the model is TRAINED.

This operation requires permissions to perform the lookoutvision:StopModel operation.

Creating a Request

data StopModel

See: newStopModel smart constructor.

Constructors

StopModel'

Fields

clientToken :: Maybe Text

ClientToken is an idempotency token that ensures a call to StopModel completes only once. You choose the value to pass. For example, An issue, such as an network outage, might prevent you from getting a response from StopModel. In this case, safely retry your call to StopModel by using the same ClientToken parameter value. An error occurs if the other input parameters are not the same as in the first request. Using a different value for ClientToken is considered a new call to StopModel. An idempotency token is active for 8 hours.

projectName :: Text

The name of the project that contains the model that you want to stop.

modelVersion :: Text

The version of the model that you want to stop.

newStopModel

Arguments

:: Text
$sel:projectName:StopModel'

-> Text
$sel:modelVersion:StopModel'

-> StopModel

Create a value of StopModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:StopModel', stopModel_clientToken - ClientToken is an idempotency token that ensures a call to StopModel completes only once. You choose the value to pass. For example, An issue, such as an network outage, might prevent you from getting a response from StopModel. In this case, safely retry your call to StopModel by using the same ClientToken parameter value. An error occurs if the other input parameters are not the same as in the first request. Using a different value for ClientToken is considered a new call to StopModel. An idempotency token is active for 8 hours.

$sel:projectName:StopModel', stopModel_projectName - The name of the project that contains the model that you want to stop.

$sel:modelVersion:StopModel', stopModel_modelVersion - The version of the model that you want to stop.

Request Lenses

stopModel_clientToken :: Lens' StopModel (Maybe Text)

ClientToken is an idempotency token that ensures a call to StopModel completes only once. You choose the value to pass. For example, An issue, such as an network outage, might prevent you from getting a response from StopModel. In this case, safely retry your call to StopModel by using the same ClientToken parameter value. An error occurs if the other input parameters are not the same as in the first request. Using a different value for ClientToken is considered a new call to StopModel. An idempotency token is active for 8 hours.

stopModel_projectName :: Lens' StopModel Text

The name of the project that contains the model that you want to stop.

stopModel_modelVersion :: Lens' StopModel Text

The version of the model that you want to stop.

Destructuring the Response

data StopModelResponse

See: newStopModelResponse smart constructor.

Constructors

StopModelResponse'

Fields

status :: Maybe ModelHostingStatus

The status of the model.

httpStatus :: Int

The response's http status code.

newStopModelResponse

Arguments

:: Int
$sel:httpStatus:StopModelResponse'

-> StopModelResponse

Create a value of StopModelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:StopModelResponse', stopModelResponse_status - The status of the model.

$sel:httpStatus:StopModelResponse', stopModelResponse_httpStatus - The response's http status code.

Response Lenses

stopModelResponse_status :: Lens' StopModelResponse (Maybe ModelHostingStatus)

The status of the model.

stopModelResponse_httpStatus :: Lens' StopModelResponse Int

The response's http status code.