Network.AWS.LookoutVision.Types.DetectAnomalyResult

Description

Documentation

data DetectAnomalyResult

The prediction results from a call to DetectAnomalies.

See: newDetectAnomalyResult smart constructor.

Constructors

DetectAnomalyResult'

Fields

isAnomalous :: Maybe Bool

True if the image contains an anomaly, otherwise false.

confidence :: Maybe Double

The confidence that Amazon Lookout for Vision has in the accuracy of the prediction.

source :: Maybe ImageSource

The source of the image that was analyzed. direct means that the images was supplied from the local computer. No other values are supported.

newDetectAnomalyResult :: DetectAnomalyResult

Create a value of DetectAnomalyResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:isAnomalous:DetectAnomalyResult', detectAnomalyResult_isAnomalous - True if the image contains an anomaly, otherwise false.

$sel:confidence:DetectAnomalyResult', detectAnomalyResult_confidence - The confidence that Amazon Lookout for Vision has in the accuracy of the prediction.

$sel:source:DetectAnomalyResult', detectAnomalyResult_source - The source of the image that was analyzed. direct means that the images was supplied from the local computer. No other values are supported.

detectAnomalyResult_isAnomalous :: Lens' DetectAnomalyResult (Maybe Bool)

True if the image contains an anomaly, otherwise false.

detectAnomalyResult_confidence :: Lens' DetectAnomalyResult (Maybe Double)

The confidence that Amazon Lookout for Vision has in the accuracy of the prediction.

detectAnomalyResult_source :: Lens' DetectAnomalyResult (Maybe ImageSource)

The source of the image that was analyzed. direct means that the images was supplied from the local computer. No other values are supported.