Network.AWS.MGN.DescribeJobs

Description

Returns a list of Jobs. Use the JobsID and fromDate and toData filters to limit which jobs are returned. The response is sorted by creationDataTime - latest date first. Jobs are normaly created by the StartTest, StartCutover, and TerminateTargetInstances APIs. Jobs are also created by DiagnosticLaunch and TerminateDiagnosticInstances, which are APIs available only to *Support* and only used in response to relevant support tickets.

This operation returns paginated results.

Creating a Request

data DescribeJobs

See: newDescribeJobs smart constructor.

Constructors

DescribeJobs'

Fields

nextToken :: Maybe Text

Request to describe Job logby next token.

maxResults :: Maybe Natural

Request to describe Job log by max results.

filters :: DescribeJobsRequestFilters

Request to describe Job log filters.

newDescribeJobs

Arguments

:: DescribeJobsRequestFilters
$sel:filters:DescribeJobs'

-> DescribeJobs

Create a value of DescribeJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeJobs', describeJobs_nextToken - Request to describe Job logby next token.

$sel:maxResults:DescribeJobs', describeJobs_maxResults - Request to describe Job log by max results.

$sel:filters:DescribeJobs', describeJobs_filters - Request to describe Job log filters.

Request Lenses

describeJobs_nextToken :: Lens' DescribeJobs (Maybe Text)

Request to describe Job logby next token.

describeJobs_maxResults :: Lens' DescribeJobs (Maybe Natural)

Request to describe Job log by max results.

describeJobs_filters :: Lens' DescribeJobs DescribeJobsRequestFilters

Request to describe Job log filters.

Destructuring the Response

data DescribeJobsResponse

See: newDescribeJobsResponse smart constructor.

Constructors

DescribeJobsResponse'

Fields

items :: Maybe [Job]

Request to describe Job log items.

nextToken :: Maybe Text

Request to describe Job response by next token.

httpStatus :: Int

The response's http status code.

newDescribeJobsResponse

Arguments

:: Int
$sel:httpStatus:DescribeJobsResponse'

-> DescribeJobsResponse

Create a value of DescribeJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:DescribeJobsResponse', describeJobsResponse_items - Request to describe Job log items.

$sel:nextToken:DescribeJobs', describeJobsResponse_nextToken - Request to describe Job response by next token.

$sel:httpStatus:DescribeJobsResponse', describeJobsResponse_httpStatus - The response's http status code.

Response Lenses

describeJobsResponse_items :: Lens' DescribeJobsResponse (Maybe [Job])

Request to describe Job log items.

describeJobsResponse_nextToken :: Lens' DescribeJobsResponse (Maybe Text)

Request to describe Job response by next token.

describeJobsResponse_httpStatus :: Lens' DescribeJobsResponse Int

The response's http status code.