Network.AWS.MGN.GetReplicationConfiguration

Description

Lists all ReplicationConfigurations, filtered by Source Server ID.

Creating a Request

data GetReplicationConfiguration

See: newGetReplicationConfiguration smart constructor.

Constructors

GetReplicationConfiguration'

Fields

sourceServerID :: Text

Request to get Replication Configuaration by Source Server ID.

newGetReplicationConfiguration

Arguments

:: Text
$sel:sourceServerID:GetReplicationConfiguration'

-> GetReplicationConfiguration

Create a value of GetReplicationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceServerID:GetReplicationConfiguration', getReplicationConfiguration_sourceServerID - Request to get Replication Configuaration by Source Server ID.

Request Lenses

getReplicationConfiguration_sourceServerID :: Lens' GetReplicationConfiguration Text

Request to get Replication Configuaration by Source Server ID.

Destructuring the Response

data ReplicationConfiguration

See: newReplicationConfiguration smart constructor.

Constructors

ReplicationConfiguration'

Fields

createPublicIP :: Maybe Bool

Replication Configuration create Public IP.

stagingAreaTags :: Maybe (Sensitive (HashMap Text Text))

Replication Configuration Staging Area tags.

stagingAreaSubnetId :: Maybe Text

Replication Configuration Staging Area subnet ID.

replicationServerInstanceType :: Maybe Text

Replication Configuration Replication Server instance type.

ebsEncryption :: Maybe ReplicationConfigurationEbsEncryption

Replication Configuration EBS encryption.

associateDefaultSecurityGroup :: Maybe Bool

Replication Configuration associate default Application Migration Service Security Group.

replicationServersSecurityGroupsIDs :: Maybe [Text]

Replication Configuration Replication Server Security Group IDs.

ebsEncryptionKeyArn :: Maybe Text

Replication Configuration EBS encryption key ARN.

name :: Maybe Text

Replication Configuration name.

defaultLargeStagingDiskType :: Maybe ReplicationConfigurationDefaultLargeStagingDiskType

Replication Configuration use default large Staging Disks.

bandwidthThrottling :: Maybe Natural

Replication Configuration set bandwidth throttling.

dataPlaneRouting :: Maybe ReplicationConfigurationDataPlaneRouting

Replication Configuration data plane routing.

sourceServerID :: Maybe Text

Replication Configuration Source Server ID.

replicatedDisks :: Maybe [ReplicationConfigurationReplicatedDisk]

Replication Configuration replicated disks.

useDedicatedReplicationServer :: Maybe Bool

Replication Configuration use Dedicated Replication Server.

newReplicationConfiguration :: ReplicationConfiguration

Create a value of ReplicationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createPublicIP:ReplicationConfiguration', replicationConfiguration_createPublicIP - Replication Configuration create Public IP.

$sel:stagingAreaTags:ReplicationConfiguration', replicationConfiguration_stagingAreaTags - Replication Configuration Staging Area tags.

$sel:stagingAreaSubnetId:ReplicationConfiguration', replicationConfiguration_stagingAreaSubnetId - Replication Configuration Staging Area subnet ID.

$sel:replicationServerInstanceType:ReplicationConfiguration', replicationConfiguration_replicationServerInstanceType - Replication Configuration Replication Server instance type.

$sel:ebsEncryption:ReplicationConfiguration', replicationConfiguration_ebsEncryption - Replication Configuration EBS encryption.

$sel:associateDefaultSecurityGroup:ReplicationConfiguration', replicationConfiguration_associateDefaultSecurityGroup - Replication Configuration associate default Application Migration Service Security Group.

$sel:replicationServersSecurityGroupsIDs:ReplicationConfiguration', replicationConfiguration_replicationServersSecurityGroupsIDs - Replication Configuration Replication Server Security Group IDs.

$sel:ebsEncryptionKeyArn:ReplicationConfiguration', replicationConfiguration_ebsEncryptionKeyArn - Replication Configuration EBS encryption key ARN.

$sel:name:ReplicationConfiguration', replicationConfiguration_name - Replication Configuration name.

$sel:defaultLargeStagingDiskType:ReplicationConfiguration', replicationConfiguration_defaultLargeStagingDiskType - Replication Configuration use default large Staging Disks.

$sel:bandwidthThrottling:ReplicationConfiguration', replicationConfiguration_bandwidthThrottling - Replication Configuration set bandwidth throttling.

$sel:dataPlaneRouting:ReplicationConfiguration', replicationConfiguration_dataPlaneRouting - Replication Configuration data plane routing.

$sel:sourceServerID:ReplicationConfiguration', replicationConfiguration_sourceServerID - Replication Configuration Source Server ID.

$sel:replicatedDisks:ReplicationConfiguration', replicationConfiguration_replicatedDisks - Replication Configuration replicated disks.

$sel:useDedicatedReplicationServer:ReplicationConfiguration', replicationConfiguration_useDedicatedReplicationServer - Replication Configuration use Dedicated Replication Server.

Response Lenses

replicationConfiguration_createPublicIP :: Lens' ReplicationConfiguration (Maybe Bool)

Replication Configuration create Public IP.

replicationConfiguration_stagingAreaTags :: Lens' ReplicationConfiguration (Maybe (HashMap Text Text))

Replication Configuration Staging Area tags.

replicationConfiguration_stagingAreaSubnetId :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration Staging Area subnet ID.

replicationConfiguration_replicationServerInstanceType :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration Replication Server instance type.

replicationConfiguration_ebsEncryption :: Lens' ReplicationConfiguration (Maybe ReplicationConfigurationEbsEncryption)

Replication Configuration EBS encryption.

replicationConfiguration_associateDefaultSecurityGroup :: Lens' ReplicationConfiguration (Maybe Bool)

Replication Configuration associate default Application Migration Service Security Group.

replicationConfiguration_replicationServersSecurityGroupsIDs :: Lens' ReplicationConfiguration (Maybe [Text])

Replication Configuration Replication Server Security Group IDs.

replicationConfiguration_ebsEncryptionKeyArn :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration EBS encryption key ARN.

replicationConfiguration_name :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration name.

replicationConfiguration_defaultLargeStagingDiskType :: Lens' ReplicationConfiguration (Maybe ReplicationConfigurationDefaultLargeStagingDiskType)

Replication Configuration use default large Staging Disks.

replicationConfiguration_bandwidthThrottling :: Lens' ReplicationConfiguration (Maybe Natural)

Replication Configuration set bandwidth throttling.

replicationConfiguration_dataPlaneRouting :: Lens' ReplicationConfiguration (Maybe ReplicationConfigurationDataPlaneRouting)

Replication Configuration data plane routing.

replicationConfiguration_sourceServerID :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration Source Server ID.

replicationConfiguration_replicatedDisks :: Lens' ReplicationConfiguration (Maybe [ReplicationConfigurationReplicatedDisk])

Replication Configuration replicated disks.

replicationConfiguration_useDedicatedReplicationServer :: Lens' ReplicationConfiguration (Maybe Bool)

Replication Configuration use Dedicated Replication Server.