Network.AWS.MGN.StartCutover

Description

Launches a Cutover Instance for specific Source Servers. This command starts a LAUNCH job whose initiatedBy property is StartCutover and changes the SourceServer.lifeCycle.state property to CUTTING_OVER.

Creating a Request

data StartCutover

See: newStartCutover smart constructor.

Constructors

StartCutover'

Fields

tags :: Maybe (Sensitive (HashMap Text Text))

Start Cutover by Tags.

sourceServerIDs :: NonEmpty Text

Start Cutover by Source Server IDs.

newStartCutover

Arguments

:: NonEmpty Text
$sel:sourceServerIDs:StartCutover'

-> StartCutover

Create a value of StartCutover with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:StartCutover', startCutover_tags - Start Cutover by Tags.

$sel:sourceServerIDs:StartCutover', startCutover_sourceServerIDs - Start Cutover by Source Server IDs.

Request Lenses

startCutover_tags :: Lens' StartCutover (Maybe (HashMap Text Text))

Start Cutover by Tags.

startCutover_sourceServerIDs :: Lens' StartCutover (NonEmpty Text)

Start Cutover by Source Server IDs.

Destructuring the Response

data StartCutoverResponse

See: newStartCutoverResponse smart constructor.

Constructors

StartCutoverResponse'

Fields

job :: Maybe Job

Start Cutover Job response.

httpStatus :: Int

The response's http status code.

newStartCutoverResponse

Arguments

:: Int
$sel:httpStatus:StartCutoverResponse'

-> StartCutoverResponse

Create a value of StartCutoverResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:job:StartCutoverResponse', startCutoverResponse_job - Start Cutover Job response.

$sel:httpStatus:StartCutoverResponse', startCutoverResponse_httpStatus - The response's http status code.

Response Lenses

startCutoverResponse_job :: Lens' StartCutoverResponse (Maybe Job)

Start Cutover Job response.

startCutoverResponse_httpStatus :: Lens' StartCutoverResponse Int

The response's http status code.