Network.AWS.MGN.TerminateTargetInstances

Description

Starts a job that terminates specific launched EC2 Test and Cutover instances. This command will not work for any Source Server with a lifecycle.state of TESTING, CUTTING_OVER, or CUTOVER.

Creating a Request

data TerminateTargetInstances

See: newTerminateTargetInstances smart constructor.

Constructors

TerminateTargetInstances'

Fields

tags :: Maybe (Sensitive (HashMap Text Text))

Terminate Target instance by Tags.

sourceServerIDs :: NonEmpty Text

Terminate Target instance by Source Server IDs.

newTerminateTargetInstances

Arguments

:: NonEmpty Text
$sel:sourceServerIDs:TerminateTargetInstances'

-> TerminateTargetInstances

Create a value of TerminateTargetInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:TerminateTargetInstances', terminateTargetInstances_tags - Terminate Target instance by Tags.

$sel:sourceServerIDs:TerminateTargetInstances', terminateTargetInstances_sourceServerIDs - Terminate Target instance by Source Server IDs.

Request Lenses

terminateTargetInstances_tags :: Lens' TerminateTargetInstances (Maybe (HashMap Text Text))

Terminate Target instance by Tags.

terminateTargetInstances_sourceServerIDs :: Lens' TerminateTargetInstances (NonEmpty Text)

Terminate Target instance by Source Server IDs.

Destructuring the Response

data TerminateTargetInstancesResponse

See: newTerminateTargetInstancesResponse smart constructor.

Constructors

TerminateTargetInstancesResponse'

Fields

job :: Maybe Job

Terminate Target instance Job response.

httpStatus :: Int

The response's http status code.

newTerminateTargetInstancesResponse

Arguments

:: Int
$sel:httpStatus:TerminateTargetInstancesResponse'

-> TerminateTargetInstancesResponse

Create a value of TerminateTargetInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:job:TerminateTargetInstancesResponse', terminateTargetInstancesResponse_job - Terminate Target instance Job response.

$sel:httpStatus:TerminateTargetInstancesResponse', terminateTargetInstancesResponse_httpStatus - The response's http status code.

Response Lenses

terminateTargetInstancesResponse_job :: Lens' TerminateTargetInstancesResponse (Maybe Job)

Terminate Target instance Job response.

terminateTargetInstancesResponse_httpStatus :: Lens' TerminateTargetInstancesResponse Int

The response's http status code.