Network.AWS.MGN.Types.DataReplicationInfo

Description

Documentation

data DataReplicationInfo

Request data replication info.

See: newDataReplicationInfo smart constructor.

Constructors

DataReplicationInfo'

Fields

dataReplicationInitiation :: Maybe DataReplicationInitiation

Request to query whether data replication has been initiated.

dataReplicationError :: Maybe DataReplicationError

Error in obtaining data replication info.

lagDuration :: Maybe Text

Request to query data replication lag durating.

dataReplicationState :: Maybe DataReplicationState

Request to query the data replication state.

replicatedDisks :: Maybe [DataReplicationInfoReplicatedDisk]

Request to query disks replicated.

etaDateTime :: Maybe Text

Request to query the time when data replication will be complete.

newDataReplicationInfo :: DataReplicationInfo

Create a value of DataReplicationInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataReplicationInitiation:DataReplicationInfo', dataReplicationInfo_dataReplicationInitiation - Request to query whether data replication has been initiated.

$sel:dataReplicationError:DataReplicationInfo', dataReplicationInfo_dataReplicationError - Error in obtaining data replication info.

$sel:lagDuration:DataReplicationInfo', dataReplicationInfo_lagDuration - Request to query data replication lag durating.

$sel:dataReplicationState:DataReplicationInfo', dataReplicationInfo_dataReplicationState - Request to query the data replication state.

$sel:replicatedDisks:DataReplicationInfo', dataReplicationInfo_replicatedDisks - Request to query disks replicated.

$sel:etaDateTime:DataReplicationInfo', dataReplicationInfo_etaDateTime - Request to query the time when data replication will be complete.

dataReplicationInfo_dataReplicationInitiation :: Lens' DataReplicationInfo (Maybe DataReplicationInitiation)

Request to query whether data replication has been initiated.

dataReplicationInfo_dataReplicationError :: Lens' DataReplicationInfo (Maybe DataReplicationError)

Error in obtaining data replication info.

dataReplicationInfo_lagDuration :: Lens' DataReplicationInfo (Maybe Text)

Request to query data replication lag durating.

dataReplicationInfo_dataReplicationState :: Lens' DataReplicationInfo (Maybe DataReplicationState)

Request to query the data replication state.

dataReplicationInfo_replicatedDisks :: Lens' DataReplicationInfo (Maybe [DataReplicationInfoReplicatedDisk])

Request to query disks replicated.

dataReplicationInfo_etaDateTime :: Lens' DataReplicationInfo (Maybe Text)

Request to query the time when data replication will be complete.