Network.AWS.MGN.Types.Job

Description

Documentation

data Job

Job.

See: newJob smart constructor.

Constructors

Job'

Fields

initiatedBy :: Maybe InitiatedBy

Job initiated by field.

status :: Maybe JobStatus

Job status.

participatingServers :: Maybe [ParticipatingServer]

Servers participating in a specific Job.

arn :: Maybe Text

the ARN of the specific Job.

creationDateTime :: Maybe Text

Job creation time.

type' :: Maybe JobType

Job type.

endDateTime :: Maybe Text

Job end time.

tags :: Maybe (Sensitive (HashMap Text Text))

Tags associated with spcific Job.

jobID :: Text

Job ID.

newJob

Arguments

:: Text
$sel:jobID:Job'

-> Job

Create a value of Job with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:initiatedBy:Job', job_initiatedBy - Job initiated by field.

$sel:status:Job', job_status - Job status.

$sel:participatingServers:Job', job_participatingServers - Servers participating in a specific Job.

$sel:arn:Job', job_arn - the ARN of the specific Job.

$sel:creationDateTime:Job', job_creationDateTime - Job creation time.

$sel:type':Job', job_type - Job type.

$sel:endDateTime:Job', job_endDateTime - Job end time.

$sel:tags:Job', job_tags - Tags associated with spcific Job.

$sel:jobID:Job', job_jobID - Job ID.

job_initiatedBy :: Lens' Job (Maybe InitiatedBy)

Job initiated by field.

job_status :: Lens' Job (Maybe JobStatus)

Job status.

job_participatingServers :: Lens' Job (Maybe [ParticipatingServer])

Servers participating in a specific Job.

job_arn :: Lens' Job (Maybe Text)

the ARN of the specific Job.

job_creationDateTime :: Lens' Job (Maybe Text)

Job creation time.

job_type :: Lens' Job (Maybe JobType)

Job type.

job_endDateTime :: Lens' Job (Maybe Text)

Job end time.

job_tags :: Lens' Job (Maybe (HashMap Text Text))

Tags associated with spcific Job.

job_jobID :: Lens' Job Text

Job ID.