Network.AWS.MGN.UpdateReplicationConfiguration

Description

Allows you to update multiple ReplicationConfigurations by Source Server ID.

Creating a Request

data UpdateReplicationConfiguration

See: newUpdateReplicationConfiguration smart constructor.

Constructors

UpdateReplicationConfiguration'

Fields

createPublicIP :: Maybe Bool

Update replication configuration create Public IP request.

stagingAreaTags :: Maybe (Sensitive (HashMap Text Text))

Update replication configuration Staging Area Tags request.

stagingAreaSubnetId :: Maybe Text

Update replication configuration Staging Area subnet request.

replicationServerInstanceType :: Maybe Text

Update replication configuration Replication Server instance type request.

ebsEncryption :: Maybe ReplicationConfigurationEbsEncryption

Update replication configuration EBS encryption request.

associateDefaultSecurityGroup :: Maybe Bool

Update replication configuration associate default Application Migration Service Security group request.

replicationServersSecurityGroupsIDs :: Maybe [Text]

Update replication configuration Replication Server Security Groups IDs request.

ebsEncryptionKeyArn :: Maybe Text

Update replication configuration EBS encryption key ARN request.

name :: Maybe Text

Update replication configuration name request.

defaultLargeStagingDiskType :: Maybe ReplicationConfigurationDefaultLargeStagingDiskType

Update replication configuration use default large Staging Disk type request.

bandwidthThrottling :: Maybe Natural

Update replication configuration bandwidth throttling request.

dataPlaneRouting :: Maybe ReplicationConfigurationDataPlaneRouting

Update replication configuration data plane routing request.

replicatedDisks :: Maybe [ReplicationConfigurationReplicatedDisk]

Update replication configuration replicated disks request.

useDedicatedReplicationServer :: Maybe Bool

Update replication configuration use dedicated Replication Server request.

sourceServerID :: Text

Update replication configuration Source Server ID request.

newUpdateReplicationConfiguration

Arguments

:: Text
$sel:sourceServerID:UpdateReplicationConfiguration'

-> UpdateReplicationConfiguration

Create a value of UpdateReplicationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createPublicIP:UpdateReplicationConfiguration', updateReplicationConfiguration_createPublicIP - Update replication configuration create Public IP request.

$sel:stagingAreaTags:UpdateReplicationConfiguration', updateReplicationConfiguration_stagingAreaTags - Update replication configuration Staging Area Tags request.

$sel:stagingAreaSubnetId:UpdateReplicationConfiguration', updateReplicationConfiguration_stagingAreaSubnetId - Update replication configuration Staging Area subnet request.

$sel:replicationServerInstanceType:UpdateReplicationConfiguration', updateReplicationConfiguration_replicationServerInstanceType - Update replication configuration Replication Server instance type request.

$sel:ebsEncryption:UpdateReplicationConfiguration', updateReplicationConfiguration_ebsEncryption - Update replication configuration EBS encryption request.

$sel:associateDefaultSecurityGroup:UpdateReplicationConfiguration', updateReplicationConfiguration_associateDefaultSecurityGroup - Update replication configuration associate default Application Migration Service Security group request.

$sel:replicationServersSecurityGroupsIDs:UpdateReplicationConfiguration', updateReplicationConfiguration_replicationServersSecurityGroupsIDs - Update replication configuration Replication Server Security Groups IDs request.

$sel:ebsEncryptionKeyArn:UpdateReplicationConfiguration', updateReplicationConfiguration_ebsEncryptionKeyArn - Update replication configuration EBS encryption key ARN request.

$sel:name:UpdateReplicationConfiguration', updateReplicationConfiguration_name - Update replication configuration name request.

$sel:defaultLargeStagingDiskType:UpdateReplicationConfiguration', updateReplicationConfiguration_defaultLargeStagingDiskType - Update replication configuration use default large Staging Disk type request.

$sel:bandwidthThrottling:UpdateReplicationConfiguration', updateReplicationConfiguration_bandwidthThrottling - Update replication configuration bandwidth throttling request.

$sel:dataPlaneRouting:UpdateReplicationConfiguration', updateReplicationConfiguration_dataPlaneRouting - Update replication configuration data plane routing request.

$sel:replicatedDisks:UpdateReplicationConfiguration', updateReplicationConfiguration_replicatedDisks - Update replication configuration replicated disks request.

$sel:useDedicatedReplicationServer:UpdateReplicationConfiguration', updateReplicationConfiguration_useDedicatedReplicationServer - Update replication configuration use dedicated Replication Server request.

$sel:sourceServerID:UpdateReplicationConfiguration', updateReplicationConfiguration_sourceServerID - Update replication configuration Source Server ID request.

Request Lenses

updateReplicationConfiguration_createPublicIP :: Lens' UpdateReplicationConfiguration (Maybe Bool)

Update replication configuration create Public IP request.

updateReplicationConfiguration_stagingAreaTags :: Lens' UpdateReplicationConfiguration (Maybe (HashMap Text Text))

Update replication configuration Staging Area Tags request.

updateReplicationConfiguration_stagingAreaSubnetId :: Lens' UpdateReplicationConfiguration (Maybe Text)

Update replication configuration Staging Area subnet request.

updateReplicationConfiguration_replicationServerInstanceType :: Lens' UpdateReplicationConfiguration (Maybe Text)

Update replication configuration Replication Server instance type request.

updateReplicationConfiguration_ebsEncryption :: Lens' UpdateReplicationConfiguration (Maybe ReplicationConfigurationEbsEncryption)

Update replication configuration EBS encryption request.

updateReplicationConfiguration_associateDefaultSecurityGroup :: Lens' UpdateReplicationConfiguration (Maybe Bool)

Update replication configuration associate default Application Migration Service Security group request.

updateReplicationConfiguration_replicationServersSecurityGroupsIDs :: Lens' UpdateReplicationConfiguration (Maybe [Text])

Update replication configuration Replication Server Security Groups IDs request.

updateReplicationConfiguration_ebsEncryptionKeyArn :: Lens' UpdateReplicationConfiguration (Maybe Text)

Update replication configuration EBS encryption key ARN request.

updateReplicationConfiguration_name :: Lens' UpdateReplicationConfiguration (Maybe Text)

Update replication configuration name request.

updateReplicationConfiguration_defaultLargeStagingDiskType :: Lens' UpdateReplicationConfiguration (Maybe ReplicationConfigurationDefaultLargeStagingDiskType)

Update replication configuration use default large Staging Disk type request.

updateReplicationConfiguration_bandwidthThrottling :: Lens' UpdateReplicationConfiguration (Maybe Natural)

Update replication configuration bandwidth throttling request.

updateReplicationConfiguration_dataPlaneRouting :: Lens' UpdateReplicationConfiguration (Maybe ReplicationConfigurationDataPlaneRouting)

Update replication configuration data plane routing request.

updateReplicationConfiguration_replicatedDisks :: Lens' UpdateReplicationConfiguration (Maybe [ReplicationConfigurationReplicatedDisk])

Update replication configuration replicated disks request.

updateReplicationConfiguration_useDedicatedReplicationServer :: Lens' UpdateReplicationConfiguration (Maybe Bool)

Update replication configuration use dedicated Replication Server request.

updateReplicationConfiguration_sourceServerID :: Lens' UpdateReplicationConfiguration Text

Update replication configuration Source Server ID request.

Destructuring the Response

data ReplicationConfiguration

See: newReplicationConfiguration smart constructor.

Constructors

ReplicationConfiguration'

Fields

createPublicIP :: Maybe Bool

Replication Configuration create Public IP.

stagingAreaTags :: Maybe (Sensitive (HashMap Text Text))

Replication Configuration Staging Area tags.

stagingAreaSubnetId :: Maybe Text

Replication Configuration Staging Area subnet ID.

replicationServerInstanceType :: Maybe Text

Replication Configuration Replication Server instance type.

ebsEncryption :: Maybe ReplicationConfigurationEbsEncryption

Replication Configuration EBS encryption.

associateDefaultSecurityGroup :: Maybe Bool

Replication Configuration associate default Application Migration Service Security Group.

replicationServersSecurityGroupsIDs :: Maybe [Text]

Replication Configuration Replication Server Security Group IDs.

ebsEncryptionKeyArn :: Maybe Text

Replication Configuration EBS encryption key ARN.

name :: Maybe Text

Replication Configuration name.

defaultLargeStagingDiskType :: Maybe ReplicationConfigurationDefaultLargeStagingDiskType

Replication Configuration use default large Staging Disks.

bandwidthThrottling :: Maybe Natural

Replication Configuration set bandwidth throttling.

dataPlaneRouting :: Maybe ReplicationConfigurationDataPlaneRouting

Replication Configuration data plane routing.

sourceServerID :: Maybe Text

Replication Configuration Source Server ID.

replicatedDisks :: Maybe [ReplicationConfigurationReplicatedDisk]

Replication Configuration replicated disks.

useDedicatedReplicationServer :: Maybe Bool

Replication Configuration use Dedicated Replication Server.

newReplicationConfiguration :: ReplicationConfiguration

Create a value of ReplicationConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createPublicIP:ReplicationConfiguration', replicationConfiguration_createPublicIP - Replication Configuration create Public IP.

$sel:stagingAreaTags:ReplicationConfiguration', replicationConfiguration_stagingAreaTags - Replication Configuration Staging Area tags.

$sel:stagingAreaSubnetId:ReplicationConfiguration', replicationConfiguration_stagingAreaSubnetId - Replication Configuration Staging Area subnet ID.

$sel:replicationServerInstanceType:ReplicationConfiguration', replicationConfiguration_replicationServerInstanceType - Replication Configuration Replication Server instance type.

$sel:ebsEncryption:ReplicationConfiguration', replicationConfiguration_ebsEncryption - Replication Configuration EBS encryption.

$sel:associateDefaultSecurityGroup:ReplicationConfiguration', replicationConfiguration_associateDefaultSecurityGroup - Replication Configuration associate default Application Migration Service Security Group.

$sel:replicationServersSecurityGroupsIDs:ReplicationConfiguration', replicationConfiguration_replicationServersSecurityGroupsIDs - Replication Configuration Replication Server Security Group IDs.

$sel:ebsEncryptionKeyArn:ReplicationConfiguration', replicationConfiguration_ebsEncryptionKeyArn - Replication Configuration EBS encryption key ARN.

$sel:name:ReplicationConfiguration', replicationConfiguration_name - Replication Configuration name.

$sel:defaultLargeStagingDiskType:ReplicationConfiguration', replicationConfiguration_defaultLargeStagingDiskType - Replication Configuration use default large Staging Disks.

$sel:bandwidthThrottling:ReplicationConfiguration', replicationConfiguration_bandwidthThrottling - Replication Configuration set bandwidth throttling.

$sel:dataPlaneRouting:ReplicationConfiguration', replicationConfiguration_dataPlaneRouting - Replication Configuration data plane routing.

$sel:sourceServerID:ReplicationConfiguration', replicationConfiguration_sourceServerID - Replication Configuration Source Server ID.

$sel:replicatedDisks:ReplicationConfiguration', replicationConfiguration_replicatedDisks - Replication Configuration replicated disks.

$sel:useDedicatedReplicationServer:ReplicationConfiguration', replicationConfiguration_useDedicatedReplicationServer - Replication Configuration use Dedicated Replication Server.

Response Lenses

replicationConfiguration_createPublicIP :: Lens' ReplicationConfiguration (Maybe Bool)

Replication Configuration create Public IP.

replicationConfiguration_stagingAreaTags :: Lens' ReplicationConfiguration (Maybe (HashMap Text Text))

Replication Configuration Staging Area tags.

replicationConfiguration_stagingAreaSubnetId :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration Staging Area subnet ID.

replicationConfiguration_replicationServerInstanceType :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration Replication Server instance type.

replicationConfiguration_ebsEncryption :: Lens' ReplicationConfiguration (Maybe ReplicationConfigurationEbsEncryption)

Replication Configuration EBS encryption.

replicationConfiguration_associateDefaultSecurityGroup :: Lens' ReplicationConfiguration (Maybe Bool)

Replication Configuration associate default Application Migration Service Security Group.

replicationConfiguration_replicationServersSecurityGroupsIDs :: Lens' ReplicationConfiguration (Maybe [Text])

Replication Configuration Replication Server Security Group IDs.

replicationConfiguration_ebsEncryptionKeyArn :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration EBS encryption key ARN.

replicationConfiguration_name :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration name.

replicationConfiguration_defaultLargeStagingDiskType :: Lens' ReplicationConfiguration (Maybe ReplicationConfigurationDefaultLargeStagingDiskType)

Replication Configuration use default large Staging Disks.

replicationConfiguration_bandwidthThrottling :: Lens' ReplicationConfiguration (Maybe Natural)

Replication Configuration set bandwidth throttling.

replicationConfiguration_dataPlaneRouting :: Lens' ReplicationConfiguration (Maybe ReplicationConfigurationDataPlaneRouting)

Replication Configuration data plane routing.

replicationConfiguration_sourceServerID :: Lens' ReplicationConfiguration (Maybe Text)

Replication Configuration Source Server ID.

replicationConfiguration_replicatedDisks :: Lens' ReplicationConfiguration (Maybe [ReplicationConfigurationReplicatedDisk])

Replication Configuration replicated disks.

replicationConfiguration_useDedicatedReplicationServer :: Lens' ReplicationConfiguration (Maybe Bool)

Replication Configuration use Dedicated Replication Server.