Network.AWS.MQ.DescribeBrokerInstanceOptions

Description

Describe available broker instance options.

Creating a Request

data DescribeBrokerInstanceOptions

See: newDescribeBrokerInstanceOptions smart constructor.

Constructors

DescribeBrokerInstanceOptions'

Fields

nextToken :: Maybe Text

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

storageType :: Maybe Text

Filter response by storage type.

maxResults :: Maybe Natural

The maximum number of brokers that Amazon MQ can return per page (20 by default). This value must be an integer from 5 to 100.

engineType :: Maybe Text

Filter response by engine type.

hostInstanceType :: Maybe Text

Filter response by host instance type.

newDescribeBrokerInstanceOptions :: DescribeBrokerInstanceOptions

Create a value of DescribeBrokerInstanceOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeBrokerInstanceOptions', describeBrokerInstanceOptions_nextToken - The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

$sel:storageType:DescribeBrokerInstanceOptions', describeBrokerInstanceOptions_storageType - Filter response by storage type.

$sel:maxResults:DescribeBrokerInstanceOptions', describeBrokerInstanceOptions_maxResults - The maximum number of brokers that Amazon MQ can return per page (20 by default). This value must be an integer from 5 to 100.

$sel:engineType:DescribeBrokerInstanceOptions', describeBrokerInstanceOptions_engineType - Filter response by engine type.

$sel:hostInstanceType:DescribeBrokerInstanceOptions', describeBrokerInstanceOptions_hostInstanceType - Filter response by host instance type.

Request Lenses

describeBrokerInstanceOptions_nextToken :: Lens' DescribeBrokerInstanceOptions (Maybe Text)

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

describeBrokerInstanceOptions_storageType :: Lens' DescribeBrokerInstanceOptions (Maybe Text)

Filter response by storage type.

describeBrokerInstanceOptions_maxResults :: Lens' DescribeBrokerInstanceOptions (Maybe Natural)

The maximum number of brokers that Amazon MQ can return per page (20 by default). This value must be an integer from 5 to 100.

describeBrokerInstanceOptions_engineType :: Lens' DescribeBrokerInstanceOptions (Maybe Text)

Filter response by engine type.

describeBrokerInstanceOptions_hostInstanceType :: Lens' DescribeBrokerInstanceOptions (Maybe Text)

Filter response by host instance type.

Destructuring the Response

data DescribeBrokerInstanceOptionsResponse

See: newDescribeBrokerInstanceOptionsResponse smart constructor.

Constructors

DescribeBrokerInstanceOptionsResponse'

Fields

nextToken :: Maybe Text

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

maxResults :: Maybe Natural

Required. The maximum number of instance options that can be returned per page (20 by default). This value must be an integer from 5 to 100.

brokerInstanceOptions :: Maybe [BrokerInstanceOption]

List of available broker instance options.

httpStatus :: Int

The response's http status code.

newDescribeBrokerInstanceOptionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeBrokerInstanceOptionsResponse'

-> DescribeBrokerInstanceOptionsResponse

Create a value of DescribeBrokerInstanceOptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeBrokerInstanceOptions', describeBrokerInstanceOptionsResponse_nextToken - The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

$sel:maxResults:DescribeBrokerInstanceOptions', describeBrokerInstanceOptionsResponse_maxResults - Required. The maximum number of instance options that can be returned per page (20 by default). This value must be an integer from 5 to 100.

$sel:brokerInstanceOptions:DescribeBrokerInstanceOptionsResponse', describeBrokerInstanceOptionsResponse_brokerInstanceOptions - List of available broker instance options.

$sel:httpStatus:DescribeBrokerInstanceOptionsResponse', describeBrokerInstanceOptionsResponse_httpStatus - The response's http status code.

Response Lenses

describeBrokerInstanceOptionsResponse_nextToken :: Lens' DescribeBrokerInstanceOptionsResponse (Maybe Text)

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

describeBrokerInstanceOptionsResponse_maxResults :: Lens' DescribeBrokerInstanceOptionsResponse (Maybe Natural)

Required. The maximum number of instance options that can be returned per page (20 by default). This value must be an integer from 5 to 100.

describeBrokerInstanceOptionsResponse_brokerInstanceOptions :: Lens' DescribeBrokerInstanceOptionsResponse (Maybe [BrokerInstanceOption])

List of available broker instance options.

describeBrokerInstanceOptionsResponse_httpStatus :: Lens' DescribeBrokerInstanceOptionsResponse Int

The response's http status code.