Network.AWS.MQ.DescribeConfigurationRevision

Description

Returns the specified configuration revision for the specified configuration.

Creating a Request

data DescribeConfigurationRevision

See: newDescribeConfigurationRevision smart constructor.

Constructors

DescribeConfigurationRevision'

Fields

configurationRevision :: Text

The revision of the configuration.

configurationId :: Text

The unique ID that Amazon MQ generates for the configuration.

newDescribeConfigurationRevision

Arguments

:: Text
$sel:configurationRevision:DescribeConfigurationRevision'

-> Text
$sel:configurationId:DescribeConfigurationRevision'

-> DescribeConfigurationRevision

Create a value of DescribeConfigurationRevision with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationRevision:DescribeConfigurationRevision', describeConfigurationRevision_configurationRevision - The revision of the configuration.

$sel:configurationId:DescribeConfigurationRevision', describeConfigurationRevision_configurationId - The unique ID that Amazon MQ generates for the configuration.

Request Lenses

describeConfigurationRevision_configurationRevision :: Lens' DescribeConfigurationRevision Text

The revision of the configuration.

describeConfigurationRevision_configurationId :: Lens' DescribeConfigurationRevision Text

The unique ID that Amazon MQ generates for the configuration.

Destructuring the Response

data DescribeConfigurationRevisionResponse

See: newDescribeConfigurationRevisionResponse smart constructor.

Constructors

DescribeConfigurationRevisionResponse'

Fields

data' :: Maybe Text

Required. The base64-encoded XML configuration.

configurationId :: Maybe Text

Required. The unique ID that Amazon MQ generates for the configuration.

description :: Maybe Text

The description of the configuration.

created :: Maybe POSIX

Required. The date and time of the configuration.

httpStatus :: Int

The response's http status code.

newDescribeConfigurationRevisionResponse

Arguments

:: Int
$sel:httpStatus:DescribeConfigurationRevisionResponse'

-> DescribeConfigurationRevisionResponse

Create a value of DescribeConfigurationRevisionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:data':DescribeConfigurationRevisionResponse', describeConfigurationRevisionResponse_data - Required. The base64-encoded XML configuration.

$sel:configurationId:DescribeConfigurationRevision', describeConfigurationRevisionResponse_configurationId - Required. The unique ID that Amazon MQ generates for the configuration.

$sel:description:DescribeConfigurationRevisionResponse', describeConfigurationRevisionResponse_description - The description of the configuration.

$sel:created:DescribeConfigurationRevisionResponse', describeConfigurationRevisionResponse_created - Required. The date and time of the configuration.

$sel:httpStatus:DescribeConfigurationRevisionResponse', describeConfigurationRevisionResponse_httpStatus - The response's http status code.

Response Lenses

describeConfigurationRevisionResponse_data :: Lens' DescribeConfigurationRevisionResponse (Maybe Text)

Required. The base64-encoded XML configuration.

describeConfigurationRevisionResponse_configurationId :: Lens' DescribeConfigurationRevisionResponse (Maybe Text)

Required. The unique ID that Amazon MQ generates for the configuration.

describeConfigurationRevisionResponse_description :: Lens' DescribeConfigurationRevisionResponse (Maybe Text)

The description of the configuration.

describeConfigurationRevisionResponse_created :: Lens' DescribeConfigurationRevisionResponse (Maybe UTCTime)

Required. The date and time of the configuration.

describeConfigurationRevisionResponse_httpStatus :: Lens' DescribeConfigurationRevisionResponse Int

The response's http status code.