Network.AWS.MQ.Lens

Description

Operations

DescribeBrokerInstanceOptions

describeBrokerInstanceOptions_nextToken :: Lens' DescribeBrokerInstanceOptions (Maybe Text)

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

describeBrokerInstanceOptions_storageType :: Lens' DescribeBrokerInstanceOptions (Maybe Text)

Filter response by storage type.

describeBrokerInstanceOptions_maxResults :: Lens' DescribeBrokerInstanceOptions (Maybe Natural)

The maximum number of brokers that Amazon MQ can return per page (20 by default). This value must be an integer from 5 to 100.

describeBrokerInstanceOptions_engineType :: Lens' DescribeBrokerInstanceOptions (Maybe Text)

Filter response by engine type.

describeBrokerInstanceOptions_hostInstanceType :: Lens' DescribeBrokerInstanceOptions (Maybe Text)

Filter response by host instance type.

describeBrokerInstanceOptionsResponse_nextToken :: Lens' DescribeBrokerInstanceOptionsResponse (Maybe Text)

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

describeBrokerInstanceOptionsResponse_maxResults :: Lens' DescribeBrokerInstanceOptionsResponse (Maybe Natural)

Required. The maximum number of instance options that can be returned per page (20 by default). This value must be an integer from 5 to 100.

describeBrokerInstanceOptionsResponse_brokerInstanceOptions :: Lens' DescribeBrokerInstanceOptionsResponse (Maybe [BrokerInstanceOption])

List of available broker instance options.

describeBrokerInstanceOptionsResponse_httpStatus :: Lens' DescribeBrokerInstanceOptionsResponse Int

The response's http status code.

CreateBroker

createBroker_encryptionOptions :: Lens' CreateBroker (Maybe EncryptionOptions)

Encryption options for the broker. Does not apply to RabbitMQ brokers.

createBroker_ldapServerMetadata :: Lens' CreateBroker (Maybe LdapServerMetadataInput)

Optional. The metadata of the LDAP server used to authenticate and authorize connections to the broker. Does not apply to RabbitMQ brokers.

createBroker_storageType :: Lens' CreateBroker (Maybe BrokerStorageType)

The broker's storage type.

createBroker_authenticationStrategy :: Lens' CreateBroker (Maybe AuthenticationStrategy)

Optional. The authentication strategy used to secure the broker. The default is SIMPLE.

createBroker_configuration :: Lens' CreateBroker (Maybe ConfigurationId)

A list of information about the configuration.

createBroker_creatorRequestId :: Lens' CreateBroker (Maybe Text)

The unique ID that the requester receives for the created broker. Amazon MQ passes your ID with the API action. Note: We recommend using a Universally Unique Identifier (UUID) for the creatorRequestId. You may omit the creatorRequestId if your application doesn't require idempotency.

createBroker_subnetIds :: Lens' CreateBroker (Maybe [Text])

The list of groups that define which subnets and IP ranges the broker can use from different Availability Zones. If you specify more than one subnet, the subnets must be in different Availability Zones. Amazon MQ will not be able to create VPC endpoints for your broker with multiple subnets in the same Availability Zone. A SINGLE_INSTANCE deployment requires one subnet (for example, the default subnet). An ACTIVE_STANDBY_MULTI_AZ Amazon MQ for ActiveMQ deployment requires two subnets. A CLUSTER_MULTI_AZ Amazon MQ for RabbitMQ deployment has no subnet requirements when deployed with public accessibility. Deployment without public accessibility requires at least one subnet.

If you specify subnets in a shared VPC for a RabbitMQ broker, the associated VPC to which the specified subnets belong must be owned by your AWS account. Amazon MQ will not be able to create VPC endpoints in VPCs that are not owned by your AWS account.

createBroker_securityGroups :: Lens' CreateBroker (Maybe [Text])

The list of rules (1 minimum, 125 maximum) that authorize connections to brokers.

createBroker_logs :: Lens' CreateBroker (Maybe Logs)

Enables Amazon CloudWatch logging for brokers.

createBroker_maintenanceWindowStartTime :: Lens' CreateBroker (Maybe WeeklyStartTime)

The parameters that determine the WeeklyStartTime.

createBroker_tags :: Lens' CreateBroker (Maybe (HashMap Text Text))

Create tags when creating the broker.

createBroker_engineVersion :: Lens' CreateBroker Text

Required. The broker engine's version. For a list of supported engine versions, see Supported engines.

createBroker_hostInstanceType :: Lens' CreateBroker Text

Required. The broker's instance type.

createBroker_autoMinorVersionUpgrade :: Lens' CreateBroker Bool

Enables automatic upgrades to new minor versions for brokers, as new versions are released and supported by Amazon MQ. Automatic upgrades occur during the scheduled maintenance window of the broker or after a manual broker reboot. Set to true by default, if no value is specified.

createBroker_users :: Lens' CreateBroker [User]

Required. The list of broker users (persons or applications) who can access queues and topics. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

Amazon MQ for RabbitMQ

When you create an Amazon MQ for RabbitMQ broker, one and only one administrative user is accepted and created when a broker is first provisioned. All subsequent broker users are created by making RabbitMQ API calls directly to brokers or via the RabbitMQ web console.

createBroker_brokerName :: Lens' CreateBroker Text

Required. The broker's name. This value must be unique in your AWS account, 1-50 characters long, must contain only letters, numbers, dashes, and underscores, and must not contain white spaces, brackets, wildcard characters, or special characters.

createBroker_deploymentMode :: Lens' CreateBroker DeploymentMode

Required. The broker's deployment mode.

createBroker_engineType :: Lens' CreateBroker EngineType

Required. The type of broker engine. Currently, Amazon MQ supports ACTIVEMQ and RABBITMQ.

createBroker_publiclyAccessible :: Lens' CreateBroker Bool

Enables connections from applications outside of the VPC that hosts the broker's subnets. Set to false by default, if no value is provided.

createBrokerResponse_brokerId :: Lens' CreateBrokerResponse (Maybe Text)

The unique ID that Amazon MQ generates for the broker.

createBrokerResponse_brokerArn :: Lens' CreateBrokerResponse (Maybe Text)

The broker's Amazon Resource Name (ARN).

createBrokerResponse_httpStatus :: Lens' CreateBrokerResponse Int

The response's http status code.

ListConfigurations

listConfigurations_nextToken :: Lens' ListConfigurations (Maybe Text)

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

listConfigurations_maxResults :: Lens' ListConfigurations (Maybe Natural)

The maximum number of brokers that Amazon MQ can return per page (20 by default). This value must be an integer from 5 to 100.

listConfigurationsResponse_nextToken :: Lens' ListConfigurationsResponse (Maybe Text)

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

listConfigurationsResponse_maxResults :: Lens' ListConfigurationsResponse (Maybe Int)

The maximum number of configurations that Amazon MQ can return per page (20 by default). This value must be an integer from 5 to 100.

listConfigurationsResponse_configurations :: Lens' ListConfigurationsResponse (Maybe [Configuration])

The list of all revisions for the specified configuration.

listConfigurationsResponse_httpStatus :: Lens' ListConfigurationsResponse Int

The response's http status code.

UpdateConfiguration

updateConfiguration_description :: Lens' UpdateConfiguration (Maybe Text)

The description of the configuration.

updateConfiguration_configurationId :: Lens' UpdateConfiguration Text

The unique ID that Amazon MQ generates for the configuration.

updateConfiguration_data :: Lens' UpdateConfiguration Text

Required. The base64-encoded XML configuration.

updateConfigurationResponse_warnings :: Lens' UpdateConfigurationResponse (Maybe [SanitizationWarning])

The list of the first 20 warnings about the configuration XML elements or attributes that were sanitized.

updateConfigurationResponse_latestRevision :: Lens' UpdateConfigurationResponse (Maybe ConfigurationRevision)

The latest revision of the configuration.

updateConfigurationResponse_arn :: Lens' UpdateConfigurationResponse (Maybe Text)

Required. The Amazon Resource Name (ARN) of the configuration.

updateConfigurationResponse_id :: Lens' UpdateConfigurationResponse (Maybe Text)

Required. The unique ID that Amazon MQ generates for the configuration.

updateConfigurationResponse_name :: Lens' UpdateConfigurationResponse (Maybe Text)

Required. The name of the configuration. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 1-150 characters long.

updateConfigurationResponse_created :: Lens' UpdateConfigurationResponse (Maybe UTCTime)

Required. The date and time of the configuration.

updateConfigurationResponse_httpStatus :: Lens' UpdateConfigurationResponse Int

The response's http status code.

DescribeBroker

describeBroker_brokerId :: Lens' DescribeBroker Text

The unique ID that Amazon MQ generates for the broker.

describeBrokerResponse_encryptionOptions :: Lens' DescribeBrokerResponse (Maybe EncryptionOptions)

Encryption options for the broker. Does not apply to RabbitMQ brokers.

describeBrokerResponse_brokerName :: Lens' DescribeBrokerResponse (Maybe Text)

The broker's name. This value must be unique in your AWS account, 1-50 characters long, must contain only letters, numbers, dashes, and underscores, and must not contain white spaces, brackets, wildcard characters, or special characters.

describeBrokerResponse_ldapServerMetadata :: Lens' DescribeBrokerResponse (Maybe LdapServerMetadataOutput)

The metadata of the LDAP server used to authenticate and authorize connections to the broker.

describeBrokerResponse_storageType :: Lens' DescribeBrokerResponse (Maybe BrokerStorageType)

The broker's storage type.

describeBrokerResponse_brokerInstances :: Lens' DescribeBrokerResponse (Maybe [BrokerInstance])

A list of information about allocated brokers.

describeBrokerResponse_pendingHostInstanceType :: Lens' DescribeBrokerResponse (Maybe Text)

The broker's host instance type to upgrade to. For a list of supported instance types, see Broker instance types.

describeBrokerResponse_brokerId :: Lens' DescribeBrokerResponse (Maybe Text)

The unique ID that Amazon MQ generates for the broker.

describeBrokerResponse_engineType :: Lens' DescribeBrokerResponse (Maybe EngineType)

The type of broker engine. Currently, Amazon MQ supports ACTIVEMQ and RABBITMQ.

describeBrokerResponse_configurations :: Lens' DescribeBrokerResponse (Maybe Configurations)

The list of all revisions for the specified configuration.

describeBrokerResponse_authenticationStrategy :: Lens' DescribeBrokerResponse (Maybe AuthenticationStrategy)

The authentication strategy used to secure the broker. The default is SIMPLE.

describeBrokerResponse_subnetIds :: Lens' DescribeBrokerResponse (Maybe [Text])

The list of groups that define which subnets and IP ranges the broker can use from different Availability Zones.

describeBrokerResponse_pendingSecurityGroups :: Lens' DescribeBrokerResponse (Maybe [Text])

The list of pending security groups to authorize connections to brokers.

describeBrokerResponse_publiclyAccessible :: Lens' DescribeBrokerResponse (Maybe Bool)

Enables connections from applications outside of the VPC that hosts the broker's subnets.

describeBrokerResponse_securityGroups :: Lens' DescribeBrokerResponse (Maybe [Text])

The list of rules (1 minimum, 125 maximum) that authorize connections to brokers.

describeBrokerResponse_logs :: Lens' DescribeBrokerResponse (Maybe LogsSummary)

The list of information about logs currently enabled and pending to be deployed for the specified broker.

describeBrokerResponse_engineVersion :: Lens' DescribeBrokerResponse (Maybe Text)

The broker engine's version. For a list of supported engine versions, see Supported engines.

describeBrokerResponse_maintenanceWindowStartTime :: Lens' DescribeBrokerResponse (Maybe WeeklyStartTime)

The parameters that determine the WeeklyStartTime.

describeBrokerResponse_pendingAuthenticationStrategy :: Lens' DescribeBrokerResponse (Maybe AuthenticationStrategy)

The authentication strategy that will be applied when the broker is rebooted. The default is SIMPLE.

describeBrokerResponse_brokerState :: Lens' DescribeBrokerResponse (Maybe BrokerState)

The broker's status.

describeBrokerResponse_tags :: Lens' DescribeBrokerResponse (Maybe (HashMap Text Text))

The list of all tags associated with this broker.

describeBrokerResponse_hostInstanceType :: Lens' DescribeBrokerResponse (Maybe Text)

The broker's instance type.

describeBrokerResponse_brokerArn :: Lens' DescribeBrokerResponse (Maybe Text)

The broker's Amazon Resource Name (ARN).

describeBrokerResponse_pendingLdapServerMetadata :: Lens' DescribeBrokerResponse (Maybe LdapServerMetadataOutput)

The metadata of the LDAP server that will be used to authenticate and authorize connections to the broker after it is rebooted.

describeBrokerResponse_pendingEngineVersion :: Lens' DescribeBrokerResponse (Maybe Text)

The broker engine version to upgrade to. For a list of supported engine versions, see Supported engines.

describeBrokerResponse_created :: Lens' DescribeBrokerResponse (Maybe UTCTime)

The time when the broker was created.

describeBrokerResponse_users :: Lens' DescribeBrokerResponse (Maybe [UserSummary])

The list of all broker usernames for the specified broker.

describeBrokerResponse_deploymentMode :: Lens' DescribeBrokerResponse (Maybe DeploymentMode)

The broker's deployment mode.

describeBrokerResponse_autoMinorVersionUpgrade :: Lens' DescribeBrokerResponse (Maybe Bool)

Enables automatic upgrades to new minor versions for brokers, as new versions are released and supported by Amazon MQ. Automatic upgrades occur during the scheduled maintenance window of the broker or after a manual broker reboot.

describeBrokerResponse_httpStatus :: Lens' DescribeBrokerResponse Int

The response's http status code.

DescribeBrokerEngineTypes

describeBrokerEngineTypes_nextToken :: Lens' DescribeBrokerEngineTypes (Maybe Text)

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

describeBrokerEngineTypes_maxResults :: Lens' DescribeBrokerEngineTypes (Maybe Natural)

The maximum number of brokers that Amazon MQ can return per page (20 by default). This value must be an integer from 5 to 100.

describeBrokerEngineTypes_engineType :: Lens' DescribeBrokerEngineTypes (Maybe Text)

Filter response by engine type.

describeBrokerEngineTypesResponse_nextToken :: Lens' DescribeBrokerEngineTypesResponse (Maybe Text)

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

describeBrokerEngineTypesResponse_brokerEngineTypes :: Lens' DescribeBrokerEngineTypesResponse (Maybe [BrokerEngineType])

List of available engine types and versions.

describeBrokerEngineTypesResponse_maxResults :: Lens' DescribeBrokerEngineTypesResponse (Maybe Natural)

Required. The maximum number of engine types that can be returned per page (20 by default). This value must be an integer from 5 to 100.

describeBrokerEngineTypesResponse_httpStatus :: Lens' DescribeBrokerEngineTypesResponse Int

The response's http status code.

DeleteTags

deleteTags_tagKeys :: Lens' DeleteTags [Text]

An array of tag keys to delete

deleteTags_resourceArn :: Lens' DeleteTags Text

The Amazon Resource Name (ARN) of the resource tag.

CreateUser

createUser_groups :: Lens' CreateUser (Maybe [Text])

The list of groups (20 maximum) to which the ActiveMQ user belongs. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

createUser_consoleAccess :: Lens' CreateUser (Maybe Bool)

Enables access to the ActiveMQ Web Console for the ActiveMQ user.

createUser_username :: Lens' CreateUser Text

The username of the ActiveMQ user. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

createUser_brokerId :: Lens' CreateUser Text

The unique ID that Amazon MQ generates for the broker.

createUser_password :: Lens' CreateUser Text

Required. The password of the user. This value must be at least 12 characters long, must contain at least 4 unique characters, and must not contain commas, colons, or equal signs (,:=).

createUserResponse_httpStatus :: Lens' CreateUserResponse Int

The response's http status code.

ListBrokers

listBrokers_nextToken :: Lens' ListBrokers (Maybe Text)

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

listBrokers_maxResults :: Lens' ListBrokers (Maybe Natural)

The maximum number of brokers that Amazon MQ can return per page (20 by default). This value must be an integer from 5 to 100.

listBrokersResponse_nextToken :: Lens' ListBrokersResponse (Maybe Text)

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

listBrokersResponse_brokerSummaries :: Lens' ListBrokersResponse (Maybe [BrokerSummary])

A list of information about all brokers.

listBrokersResponse_httpStatus :: Lens' ListBrokersResponse Int

The response's http status code.

RebootBroker

rebootBroker_brokerId :: Lens' RebootBroker Text

The unique ID that Amazon MQ generates for the broker.

rebootBrokerResponse_httpStatus :: Lens' RebootBrokerResponse Int

The response's http status code.

UpdateBroker

updateBroker_ldapServerMetadata :: Lens' UpdateBroker (Maybe LdapServerMetadataInput)

Optional. The metadata of the LDAP server used to authenticate and authorize connections to the broker. Does not apply to RabbitMQ brokers.

updateBroker_authenticationStrategy :: Lens' UpdateBroker (Maybe AuthenticationStrategy)

Optional. The authentication strategy used to secure the broker. The default is SIMPLE.

updateBroker_configuration :: Lens' UpdateBroker (Maybe ConfigurationId)

A list of information about the configuration.

updateBroker_securityGroups :: Lens' UpdateBroker (Maybe [Text])

The list of security groups (1 minimum, 5 maximum) that authorizes connections to brokers.

updateBroker_logs :: Lens' UpdateBroker (Maybe Logs)

Enables Amazon CloudWatch logging for brokers.

updateBroker_engineVersion :: Lens' UpdateBroker (Maybe Text)

The broker engine version. For a list of supported engine versions, see Supported engines.

updateBroker_maintenanceWindowStartTime :: Lens' UpdateBroker (Maybe WeeklyStartTime)

The parameters that determine the WeeklyStartTime.

updateBroker_hostInstanceType :: Lens' UpdateBroker (Maybe Text)

The broker's host instance type to upgrade to. For a list of supported instance types, see Broker instance types.

updateBroker_autoMinorVersionUpgrade :: Lens' UpdateBroker (Maybe Bool)

Enables automatic upgrades to new minor versions for brokers, as new versions are released and supported by Amazon MQ. Automatic upgrades occur during the scheduled maintenance window of the broker or after a manual broker reboot.

updateBroker_brokerId :: Lens' UpdateBroker Text

The unique ID that Amazon MQ generates for the broker.

updateBrokerResponse_ldapServerMetadata :: Lens' UpdateBrokerResponse (Maybe LdapServerMetadataOutput)

Optional. The metadata of the LDAP server used to authenticate and authorize connections to the broker. Does not apply to RabbitMQ brokers.

updateBrokerResponse_brokerId :: Lens' UpdateBrokerResponse (Maybe Text)

Required. The unique ID that Amazon MQ generates for the broker.

updateBrokerResponse_authenticationStrategy :: Lens' UpdateBrokerResponse (Maybe AuthenticationStrategy)

Optional. The authentication strategy used to secure the broker. The default is SIMPLE.

updateBrokerResponse_configuration :: Lens' UpdateBrokerResponse (Maybe ConfigurationId)

The ID of the updated configuration.

updateBrokerResponse_securityGroups :: Lens' UpdateBrokerResponse (Maybe [Text])

The list of security groups (1 minimum, 5 maximum) that authorizes connections to brokers.

updateBrokerResponse_logs :: Lens' UpdateBrokerResponse (Maybe Logs)

The list of information about logs to be enabled for the specified broker.

updateBrokerResponse_engineVersion :: Lens' UpdateBrokerResponse (Maybe Text)

The broker engine version to upgrade to. For a list of supported engine versions, see Supported engines.

updateBrokerResponse_maintenanceWindowStartTime :: Lens' UpdateBrokerResponse (Maybe WeeklyStartTime)

The parameters that determine the WeeklyStartTime.

updateBrokerResponse_hostInstanceType :: Lens' UpdateBrokerResponse (Maybe Text)

The broker's host instance type to upgrade to. For a list of supported instance types, see Broker instance types.

updateBrokerResponse_autoMinorVersionUpgrade :: Lens' UpdateBrokerResponse (Maybe Bool)

The new boolean value that specifies whether broker engines automatically upgrade to new minor versions as new versions are released and supported by Amazon MQ.

updateBrokerResponse_httpStatus :: Lens' UpdateBrokerResponse Int

The response's http status code.

DeleteBroker

deleteBroker_brokerId :: Lens' DeleteBroker Text

The unique ID that Amazon MQ generates for the broker.

deleteBrokerResponse_brokerId :: Lens' DeleteBrokerResponse (Maybe Text)

The unique ID that Amazon MQ generates for the broker.

deleteBrokerResponse_httpStatus :: Lens' DeleteBrokerResponse Int

The response's http status code.

ListConfigurationRevisions

listConfigurationRevisions_nextToken :: Lens' ListConfigurationRevisions (Maybe Text)

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

listConfigurationRevisions_maxResults :: Lens' ListConfigurationRevisions (Maybe Natural)

The maximum number of brokers that Amazon MQ can return per page (20 by default). This value must be an integer from 5 to 100.

listConfigurationRevisions_configurationId :: Lens' ListConfigurationRevisions Text

The unique ID that Amazon MQ generates for the configuration.

listConfigurationRevisionsResponse_nextToken :: Lens' ListConfigurationRevisionsResponse (Maybe Text)

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

listConfigurationRevisionsResponse_maxResults :: Lens' ListConfigurationRevisionsResponse (Maybe Int)

The maximum number of configuration revisions that can be returned per page (20 by default). This value must be an integer from 5 to 100.

listConfigurationRevisionsResponse_revisions :: Lens' ListConfigurationRevisionsResponse (Maybe [ConfigurationRevision])

The list of all revisions for the specified configuration.

listConfigurationRevisionsResponse_configurationId :: Lens' ListConfigurationRevisionsResponse (Maybe Text)

The unique ID that Amazon MQ generates for the configuration.

listConfigurationRevisionsResponse_httpStatus :: Lens' ListConfigurationRevisionsResponse Int

The response's http status code.

CreateConfiguration

createConfiguration_authenticationStrategy :: Lens' CreateConfiguration (Maybe AuthenticationStrategy)

Optional. The authentication strategy associated with the configuration. The default is SIMPLE.

createConfiguration_tags :: Lens' CreateConfiguration (Maybe (HashMap Text Text))

Create tags when creating the configuration.

createConfiguration_engineVersion :: Lens' CreateConfiguration Text

Required. The broker engine's version. For a list of supported engine versions, see Supported engines.

createConfiguration_engineType :: Lens' CreateConfiguration EngineType

Required. The type of broker engine. Currently, Amazon MQ supports ACTIVEMQ and RABBITMQ.

createConfiguration_name :: Lens' CreateConfiguration Text

Required. The name of the configuration. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 1-150 characters long.

createConfigurationResponse_authenticationStrategy :: Lens' CreateConfigurationResponse (Maybe AuthenticationStrategy)

Optional. The authentication strategy associated with the configuration. The default is SIMPLE.

createConfigurationResponse_latestRevision :: Lens' CreateConfigurationResponse (Maybe ConfigurationRevision)

The latest revision of the configuration.

createConfigurationResponse_arn :: Lens' CreateConfigurationResponse (Maybe Text)

Required. The Amazon Resource Name (ARN) of the configuration.

createConfigurationResponse_id :: Lens' CreateConfigurationResponse (Maybe Text)

Required. The unique ID that Amazon MQ generates for the configuration.

createConfigurationResponse_name :: Lens' CreateConfigurationResponse (Maybe Text)

Required. The name of the configuration. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 1-150 characters long.

createConfigurationResponse_created :: Lens' CreateConfigurationResponse (Maybe UTCTime)

Required. The date and time of the configuration.

createConfigurationResponse_httpStatus :: Lens' CreateConfigurationResponse Int

The response's http status code.

DescribeUser

describeUser_username :: Lens' DescribeUser Text

The username of the ActiveMQ user. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

describeUser_brokerId :: Lens' DescribeUser Text

The unique ID that Amazon MQ generates for the broker.

describeUserResponse_groups :: Lens' DescribeUserResponse (Maybe [Text])

The list of groups (20 maximum) to which the ActiveMQ user belongs. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

describeUserResponse_brokerId :: Lens' DescribeUserResponse (Maybe Text)

Required. The unique ID that Amazon MQ generates for the broker.

describeUserResponse_pending :: Lens' DescribeUserResponse (Maybe UserPendingChanges)

The status of the changes pending for the ActiveMQ user.

describeUserResponse_username :: Lens' DescribeUserResponse (Maybe Text)

Required. The username of the ActiveMQ user. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

describeUserResponse_consoleAccess :: Lens' DescribeUserResponse (Maybe Bool)

Enables access to the the ActiveMQ Web Console for the ActiveMQ user.

describeUserResponse_httpStatus :: Lens' DescribeUserResponse Int

The response's http status code.

DescribeConfigurationRevision

describeConfigurationRevision_configurationRevision :: Lens' DescribeConfigurationRevision Text

The revision of the configuration.

describeConfigurationRevision_configurationId :: Lens' DescribeConfigurationRevision Text

The unique ID that Amazon MQ generates for the configuration.

describeConfigurationRevisionResponse_data :: Lens' DescribeConfigurationRevisionResponse (Maybe Text)

Required. The base64-encoded XML configuration.

describeConfigurationRevisionResponse_configurationId :: Lens' DescribeConfigurationRevisionResponse (Maybe Text)

Required. The unique ID that Amazon MQ generates for the configuration.

describeConfigurationRevisionResponse_description :: Lens' DescribeConfigurationRevisionResponse (Maybe Text)

The description of the configuration.

describeConfigurationRevisionResponse_created :: Lens' DescribeConfigurationRevisionResponse (Maybe UTCTime)

Required. The date and time of the configuration.

describeConfigurationRevisionResponse_httpStatus :: Lens' DescribeConfigurationRevisionResponse Int

The response's http status code.

ListTags

listTags_resourceArn :: Lens' ListTags Text

The Amazon Resource Name (ARN) of the resource tag.

listTagsResponse_tags :: Lens' ListTagsResponse (Maybe (HashMap Text Text))

The key-value pair for the resource tag.

listTagsResponse_httpStatus :: Lens' ListTagsResponse Int

The response's http status code.

UpdateUser

updateUser_groups :: Lens' UpdateUser (Maybe [Text])

The list of groups (20 maximum) to which the ActiveMQ user belongs. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

updateUser_password :: Lens' UpdateUser (Maybe Text)

The password of the user. This value must be at least 12 characters long, must contain at least 4 unique characters, and must not contain commas, colons, or equal signs (,:=).

updateUser_consoleAccess :: Lens' UpdateUser (Maybe Bool)

Enables access to the the ActiveMQ Web Console for the ActiveMQ user.

updateUser_username :: Lens' UpdateUser Text

The username of the ActiveMQ user. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

updateUser_brokerId :: Lens' UpdateUser Text

The unique ID that Amazon MQ generates for the broker.

updateUserResponse_httpStatus :: Lens' UpdateUserResponse Int

The response's http status code.

ListUsers

listUsers_nextToken :: Lens' ListUsers (Maybe Text)

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

listUsers_maxResults :: Lens' ListUsers (Maybe Natural)

The maximum number of brokers that Amazon MQ can return per page (20 by default). This value must be an integer from 5 to 100.

listUsers_brokerId :: Lens' ListUsers Text

The unique ID that Amazon MQ generates for the broker.

listUsersResponse_nextToken :: Lens' ListUsersResponse (Maybe Text)

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

listUsersResponse_maxResults :: Lens' ListUsersResponse (Maybe Natural)

Required. The maximum number of ActiveMQ users that can be returned per page (20 by default). This value must be an integer from 5 to 100.

listUsersResponse_brokerId :: Lens' ListUsersResponse (Maybe Text)

Required. The unique ID that Amazon MQ generates for the broker.

listUsersResponse_users :: Lens' ListUsersResponse (Maybe [UserSummary])

Required. The list of all ActiveMQ usernames for the specified broker. Does not apply to RabbitMQ brokers.

listUsersResponse_httpStatus :: Lens' ListUsersResponse Int

The response's http status code.

DeleteUser

deleteUser_username :: Lens' DeleteUser Text

The username of the ActiveMQ user. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

deleteUser_brokerId :: Lens' DeleteUser Text

The unique ID that Amazon MQ generates for the broker.

deleteUserResponse_httpStatus :: Lens' DeleteUserResponse Int

The response's http status code.

DescribeConfiguration

describeConfiguration_configurationId :: Lens' DescribeConfiguration Text

The unique ID that Amazon MQ generates for the configuration.

describeConfigurationResponse_engineType :: Lens' DescribeConfigurationResponse (Maybe EngineType)

Required. The type of broker engine. Currently, Amazon MQ supports ACTIVEMQ and RABBITMQ.

describeConfigurationResponse_authenticationStrategy :: Lens' DescribeConfigurationResponse (Maybe AuthenticationStrategy)

Optional. The authentication strategy associated with the configuration. The default is SIMPLE.

describeConfigurationResponse_latestRevision :: Lens' DescribeConfigurationResponse (Maybe ConfigurationRevision)

Required. The latest revision of the configuration.

describeConfigurationResponse_arn :: Lens' DescribeConfigurationResponse (Maybe Text)

Required. The ARN of the configuration.

describeConfigurationResponse_id :: Lens' DescribeConfigurationResponse (Maybe Text)

Required. The unique ID that Amazon MQ generates for the configuration.

describeConfigurationResponse_name :: Lens' DescribeConfigurationResponse (Maybe Text)

Required. The name of the configuration. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 1-150 characters long.

describeConfigurationResponse_engineVersion :: Lens' DescribeConfigurationResponse (Maybe Text)

Required. The broker engine's version. For a list of supported engine versions, see, Supported engines.

describeConfigurationResponse_tags :: Lens' DescribeConfigurationResponse (Maybe (HashMap Text Text))

The list of all tags associated with this configuration.

describeConfigurationResponse_description :: Lens' DescribeConfigurationResponse (Maybe Text)

Required. The description of the configuration.

describeConfigurationResponse_created :: Lens' DescribeConfigurationResponse (Maybe UTCTime)

Required. The date and time of the configuration revision.

describeConfigurationResponse_httpStatus :: Lens' DescribeConfigurationResponse Int

The response's http status code.

CreateTags

createTags_tags :: Lens' CreateTags (Maybe (HashMap Text Text))

The key-value pair for the resource tag.

createTags_resourceArn :: Lens' CreateTags Text

The Amazon Resource Name (ARN) of the resource tag.

Types

AvailabilityZone

availabilityZone_name :: Lens' AvailabilityZone (Maybe Text)

Id for the availability zone.

BrokerEngineType

brokerEngineType_engineType :: Lens' BrokerEngineType (Maybe EngineType)

The broker's engine type.

brokerEngineType_engineVersions :: Lens' BrokerEngineType (Maybe [EngineVersion])

The list of engine versions.

BrokerInstance

brokerInstance_endpoints :: Lens' BrokerInstance (Maybe [Text])

The broker's wire-level protocol endpoints.

brokerInstance_ipAddress :: Lens' BrokerInstance (Maybe Text)

The IP address of the Elastic Network Interface (ENI) attached to the broker. Does not apply to RabbitMQ brokers.

brokerInstance_consoleURL :: Lens' BrokerInstance (Maybe Text)

The brokers web console URL.

BrokerInstanceOption

brokerInstanceOption_availabilityZones :: Lens' BrokerInstanceOption (Maybe [AvailabilityZone])

The list of available az.

brokerInstanceOption_storageType :: Lens' BrokerInstanceOption (Maybe BrokerStorageType)

The broker's storage type.

brokerInstanceOption_engineType :: Lens' BrokerInstanceOption (Maybe EngineType)

The broker's engine type.

brokerInstanceOption_supportedDeploymentModes :: Lens' BrokerInstanceOption (Maybe [DeploymentMode])

The list of supported deployment modes.

brokerInstanceOption_supportedEngineVersions :: Lens' BrokerInstanceOption (Maybe [Text])

The list of supported engine versions.

brokerInstanceOption_hostInstanceType :: Lens' BrokerInstanceOption (Maybe Text)

The broker's instance type.

BrokerSummary

brokerSummary_brokerName :: Lens' BrokerSummary (Maybe Text)

The broker's name. This value is unique in your AWS account, 1-50 characters long, and containing only letters, numbers, dashes, and underscores, and must not contain white spaces, brackets, wildcard characters, or special characters.

brokerSummary_brokerId :: Lens' BrokerSummary (Maybe Text)

The unique ID that Amazon MQ generates for the broker.

brokerSummary_brokerState :: Lens' BrokerSummary (Maybe BrokerState)

The broker's status.

brokerSummary_hostInstanceType :: Lens' BrokerSummary (Maybe Text)

The broker's instance type.

brokerSummary_brokerArn :: Lens' BrokerSummary (Maybe Text)

The broker's Amazon Resource Name (ARN).

brokerSummary_created :: Lens' BrokerSummary (Maybe UTCTime)

The time when the broker was created.

brokerSummary_deploymentMode :: Lens' BrokerSummary DeploymentMode

The broker's deployment mode.

brokerSummary_engineType :: Lens' BrokerSummary EngineType

The type of broker engine.

Configuration

configuration_tags :: Lens' Configuration (Maybe (HashMap Text Text))

The list of all tags associated with this configuration.

configuration_description :: Lens' Configuration Text

Required. The description of the configuration.

configuration_engineVersion :: Lens' Configuration Text

Required. The broker engine's version. For a list of supported engine versions, see, Supported engines.

configuration_latestRevision :: Lens' Configuration ConfigurationRevision

Required. The latest revision of the configuration.

configuration_authenticationStrategy :: Lens' Configuration AuthenticationStrategy

Optional. The authentication strategy associated with the configuration. The default is SIMPLE.

configuration_engineType :: Lens' Configuration EngineType

Required. The type of broker engine. Currently, Amazon MQ supports ACTIVEMQ and RABBITMQ.

configuration_id :: Lens' Configuration Text

Required. The unique ID that Amazon MQ generates for the configuration.

configuration_arn :: Lens' Configuration Text

Required. The ARN of the configuration.

configuration_name :: Lens' Configuration Text

Required. The name of the configuration. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 1-150 characters long.

configuration_created :: Lens' Configuration UTCTime

Required. The date and time of the configuration revision.

ConfigurationId

configurationId_revision :: Lens' ConfigurationId (Maybe Int)

The revision number of the configuration.

configurationId_id :: Lens' ConfigurationId Text

Required. The unique ID that Amazon MQ generates for the configuration.

ConfigurationRevision

configurationRevision_description :: Lens' ConfigurationRevision (Maybe Text)

The description of the configuration revision.

configurationRevision_revision :: Lens' ConfigurationRevision Int

Required. The revision number of the configuration.

configurationRevision_created :: Lens' ConfigurationRevision UTCTime

Required. The date and time of the configuration revision.

Configurations

configurations_pending :: Lens' Configurations (Maybe ConfigurationId)

The broker's pending configuration.

configurations_current :: Lens' Configurations (Maybe ConfigurationId)

The broker's current configuration.

configurations_history :: Lens' Configurations (Maybe [ConfigurationId])

The history of configurations applied to the broker.

EncryptionOptions

encryptionOptions_kmsKeyId :: Lens' EncryptionOptions (Maybe Text)

The customer master key (CMK) to use for the AWS Key Management Service (KMS). This key is used to encrypt your data at rest. If not provided, Amazon MQ will use a default CMK to encrypt your data.

encryptionOptions_useAwsOwnedKey :: Lens' EncryptionOptions Bool

Enables the use of an AWS owned CMK using AWS Key Management Service (KMS). Set to true by default, if no value is provided, for example, for RabbitMQ brokers.

EngineVersion

engineVersion_name :: Lens' EngineVersion (Maybe Text)

Id for the version.

LdapServerMetadataInput

ldapServerMetadataInput_roleName :: Lens' LdapServerMetadataInput (Maybe Text)

Specifies the LDAP attribute that identifies the group name attribute in the object returned from the group membership query.

ldapServerMetadataInput_userSearchSubtree :: Lens' LdapServerMetadataInput (Maybe Bool)

The directory search scope for the user. If set to true, scope is to search the entire subtree.

ldapServerMetadataInput_userRoleName :: Lens' LdapServerMetadataInput (Maybe Text)

Specifies the name of the LDAP attribute for the user group membership.

ldapServerMetadataInput_roleSearchSubtree :: Lens' LdapServerMetadataInput (Maybe Bool)

The directory search scope for the role. If set to true, scope is to search the entire subtree.

ldapServerMetadataInput_hosts :: Lens' LdapServerMetadataInput [Text]

Specifies the location of the LDAP server such as AWS Directory Service for Microsoft Active Directory . Optional failover server.

ldapServerMetadataInput_userSearchMatching :: Lens' LdapServerMetadataInput Text

The LDAP search filter used to find users within the userBase. The client's username is substituted into the {0} placeholder in the search filter. For example, if this option is set to (uid={0}) and the received username is janedoe, the search filter becomes (uid=janedoe) after string substitution. It will result in matching an entry like uid=janedoe, ou=Users,ou=corp, dc=corp, dc=example, dc=com.

ldapServerMetadataInput_userBase :: Lens' LdapServerMetadataInput Text

Select a particular subtree of the directory information tree (DIT) to search for user entries. The subtree is specified by a DN, which specifies the base node of the subtree. For example, by setting this option to ou=Users,ou=corp, dc=corp, dc=example, dc=com, the search for user entries is restricted to the subtree beneath ou=Users, ou=corp, dc=corp, dc=example, dc=com.

ldapServerMetadataInput_roleSearchMatching :: Lens' LdapServerMetadataInput Text

The LDAP search filter used to find roles within the roleBase. The distinguished name of the user matched by userSearchMatching is substituted into the {0} placeholder in the search filter. The client's username is substituted into the {1} placeholder. For example, if you set this option to (member=uid={1})for the user janedoe, the search filter becomes (member=uid=janedoe) after string substitution. It matches all role entries that have a member attribute equal to uid=janedoe under the subtree selected by the roleBase.

ldapServerMetadataInput_serviceAccountUsername :: Lens' LdapServerMetadataInput Text

Service account username. A service account is an account in your LDAP server that has access to initiate a connection. For example, cn=admin,dc=corp, dc=example, dc=com.

ldapServerMetadataInput_roleBase :: Lens' LdapServerMetadataInput Text

The distinguished name of the node in the directory information tree (DIT) to search for roles or groups. For example, ou=group, ou=corp, dc=corp, dc=example, dc=com.

ldapServerMetadataInput_serviceAccountPassword :: Lens' LdapServerMetadataInput Text

Service account password. A service account is an account in your LDAP server that has access to initiate a connection. For example, cn=admin,dc=corp, dc=example, dc=com.

LdapServerMetadataOutput

ldapServerMetadataOutput_roleName :: Lens' LdapServerMetadataOutput (Maybe Text)

Specifies the LDAP attribute that identifies the group name attribute in the object returned from the group membership query.

ldapServerMetadataOutput_userSearchSubtree :: Lens' LdapServerMetadataOutput (Maybe Bool)

The directory search scope for the user. If set to true, scope is to search the entire subtree.

ldapServerMetadataOutput_userRoleName :: Lens' LdapServerMetadataOutput (Maybe Text)

Specifies the name of the LDAP attribute for the user group membership.

ldapServerMetadataOutput_roleSearchSubtree :: Lens' LdapServerMetadataOutput (Maybe Bool)

The directory search scope for the role. If set to true, scope is to search the entire subtree.

ldapServerMetadataOutput_hosts :: Lens' LdapServerMetadataOutput [Text]

Specifies the location of the LDAP server such as AWS Directory Service for Microsoft Active Directory . Optional failover server.

ldapServerMetadataOutput_userSearchMatching :: Lens' LdapServerMetadataOutput Text

The LDAP search filter used to find users within the userBase. The client's username is substituted into the {0} placeholder in the search filter. For example, if this option is set to (uid={0}) and the received username is janedoe, the search filter becomes (uid=janedoe) after string substitution. It will result in matching an entry like uid=janedoe, ou=Users,ou=corp, dc=corp, dc=example, dc=com.

ldapServerMetadataOutput_userBase :: Lens' LdapServerMetadataOutput Text

Select a particular subtree of the directory information tree (DIT) to search for user entries. The subtree is specified by a DN, which specifies the base node of the subtree. For example, by setting this option to ou=Users,ou=corp, dc=corp, dc=example, dc=com, the search for user entries is restricted to the subtree beneath ou=Users, ou=corp, dc=corp, dc=example, dc=com.

ldapServerMetadataOutput_roleSearchMatching :: Lens' LdapServerMetadataOutput Text

The LDAP search filter used to find roles within the roleBase. The distinguished name of the user matched by userSearchMatching is substituted into the {0} placeholder in the search filter. The client's username is substituted into the {1} placeholder. For example, if you set this option to (member=uid={1})for the user janedoe, the search filter becomes (member=uid=janedoe) after string substitution. It matches all role entries that have a member attribute equal to uid=janedoe under the subtree selected by the roleBase.

ldapServerMetadataOutput_serviceAccountUsername :: Lens' LdapServerMetadataOutput Text

Service account username. A service account is an account in your LDAP server that has access to initiate a connection. For example, cn=admin,dc=corp, dc=example, dc=com.

ldapServerMetadataOutput_roleBase :: Lens' LdapServerMetadataOutput Text

The distinguished name of the node in the directory information tree (DIT) to search for roles or groups. For example, ou=group, ou=corp, dc=corp, dc=example, dc=com.

Logs

logs_general :: Lens' Logs (Maybe Bool)

Enables general logging.

logs_audit :: Lens' Logs (Maybe Bool)

Enables audit logging. Every user management action made using JMX or the ActiveMQ Web Console is logged. Does not apply to RabbitMQ brokers.

LogsSummary

logsSummary_audit :: Lens' LogsSummary (Maybe Bool)

Enables audit logging. Every user management action made using JMX or the ActiveMQ Web Console is logged.

logsSummary_pending :: Lens' LogsSummary (Maybe PendingLogs)

The list of information about logs pending to be deployed for the specified broker.

logsSummary_auditLogGroup :: Lens' LogsSummary (Maybe Text)

The location of the CloudWatch Logs log group where audit logs are sent.

logsSummary_generalLogGroup :: Lens' LogsSummary Text

The location of the CloudWatch Logs log group where general logs are sent.

logsSummary_general :: Lens' LogsSummary Bool

Enables general logging.

PendingLogs

pendingLogs_general :: Lens' PendingLogs (Maybe Bool)

Enables general logging.

pendingLogs_audit :: Lens' PendingLogs (Maybe Bool)

Enables audit logging. Every user management action made using JMX or the ActiveMQ Web Console is logged.

SanitizationWarning

sanitizationWarning_elementName :: Lens' SanitizationWarning (Maybe Text)

The name of the XML element that has been sanitized.

sanitizationWarning_attributeName :: Lens' SanitizationWarning (Maybe Text)

The name of the XML attribute that has been sanitized.

sanitizationWarning_reason :: Lens' SanitizationWarning SanitizationWarningReason

Required. The reason for which the XML elements or attributes were sanitized.

User

user_groups :: Lens' User (Maybe [Text])

The list of groups (20 maximum) to which the ActiveMQ user belongs. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long. Does not apply to RabbitMQ brokers.

user_consoleAccess :: Lens' User (Maybe Bool)

Enables access to the ActiveMQ Web Console for the ActiveMQ user. Does not apply to RabbitMQ brokers.

user_username :: Lens' User Text

important>

Amazon MQ for ActiveMQ

For ActiveMQ brokers, this value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

/important>

Amazon MQ for RabbitMQ

For RabbitMQ brokers, this value can contain only alphanumeric characters, dashes, periods, underscores (- . _). This value must not contain a tilde (~) character. Amazon MQ prohibts using guest as a valid usename. This value must be 2-100 characters long.

user_password :: Lens' User Text

Required. The password of the user. This value must be at least 12 characters long, must contain at least 4 unique characters, and must not contain commas, colons, or equal signs (,:=).

UserPendingChanges

userPendingChanges_groups :: Lens' UserPendingChanges (Maybe [Text])

The list of groups (20 maximum) to which the ActiveMQ user belongs. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

userPendingChanges_consoleAccess :: Lens' UserPendingChanges (Maybe Bool)

Enables access to the the ActiveMQ Web Console for the ActiveMQ user.

userPendingChanges_pendingChange :: Lens' UserPendingChanges ChangeType

Required. The type of change pending for the ActiveMQ user.

UserSummary

userSummary_pendingChange :: Lens' UserSummary (Maybe ChangeType)

The type of change pending for the broker user.

userSummary_username :: Lens' UserSummary Text

Required. The username of the broker user. This value can contain only alphanumeric characters, dashes, periods, underscores, and tildes (- . _ ~). This value must be 2-100 characters long.

WeeklyStartTime

weeklyStartTime_timeZone :: Lens' WeeklyStartTime (Maybe Text)

The time zone, UTC by default, in either the Country/City format, or the UTC offset format.

weeklyStartTime_timeOfDay :: Lens' WeeklyStartTime Text

Required. The time, in 24-hour format.

weeklyStartTime_dayOfWeek :: Lens' WeeklyStartTime DayOfWeek

Required. The day of the week.