Network.AWS.MQ.ListBrokers

Description

Returns a list of all brokers.

This operation returns paginated results.

Creating a Request

data ListBrokers

See: newListBrokers smart constructor.

Constructors

ListBrokers'

Fields

nextToken :: Maybe Text

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

maxResults :: Maybe Natural

The maximum number of brokers that Amazon MQ can return per page (20 by default). This value must be an integer from 5 to 100.

newListBrokers :: ListBrokers

Create a value of ListBrokers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListBrokers', listBrokers_nextToken - The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

$sel:maxResults:ListBrokers', listBrokers_maxResults - The maximum number of brokers that Amazon MQ can return per page (20 by default). This value must be an integer from 5 to 100.

Request Lenses

listBrokers_nextToken :: Lens' ListBrokers (Maybe Text)

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

listBrokers_maxResults :: Lens' ListBrokers (Maybe Natural)

The maximum number of brokers that Amazon MQ can return per page (20 by default). This value must be an integer from 5 to 100.

Destructuring the Response

data ListBrokersResponse

See: newListBrokersResponse smart constructor.

Constructors

ListBrokersResponse'

Fields

nextToken :: Maybe Text

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

brokerSummaries :: Maybe [BrokerSummary]

A list of information about all brokers.

httpStatus :: Int

The response's http status code.

newListBrokersResponse

Arguments

:: Int
$sel:httpStatus:ListBrokersResponse'

-> ListBrokersResponse

Create a value of ListBrokersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListBrokers', listBrokersResponse_nextToken - The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

$sel:brokerSummaries:ListBrokersResponse', listBrokersResponse_brokerSummaries - A list of information about all brokers.

$sel:httpStatus:ListBrokersResponse', listBrokersResponse_httpStatus - The response's http status code.

Response Lenses

listBrokersResponse_nextToken :: Lens' ListBrokersResponse (Maybe Text)

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

listBrokersResponse_brokerSummaries :: Lens' ListBrokersResponse (Maybe [BrokerSummary])

A list of information about all brokers.

listBrokersResponse_httpStatus :: Lens' ListBrokersResponse Int

The response's http status code.