Network.AWS.MQ.ListUsers

Description

Returns a list of all ActiveMQ users.

Creating a Request

data ListUsers

See: newListUsers smart constructor.

Constructors

ListUsers'

Fields

nextToken :: Maybe Text

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

maxResults :: Maybe Natural

The maximum number of brokers that Amazon MQ can return per page (20 by default). This value must be an integer from 5 to 100.

brokerId :: Text

The unique ID that Amazon MQ generates for the broker.

newListUsers

Arguments

:: Text
$sel:brokerId:ListUsers'

-> ListUsers

Create a value of ListUsers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListUsers', listUsers_nextToken - The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

$sel:maxResults:ListUsers', listUsers_maxResults - The maximum number of brokers that Amazon MQ can return per page (20 by default). This value must be an integer from 5 to 100.

$sel:brokerId:ListUsers', listUsers_brokerId - The unique ID that Amazon MQ generates for the broker.

Request Lenses

listUsers_nextToken :: Lens' ListUsers (Maybe Text)

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

listUsers_maxResults :: Lens' ListUsers (Maybe Natural)

The maximum number of brokers that Amazon MQ can return per page (20 by default). This value must be an integer from 5 to 100.

listUsers_brokerId :: Lens' ListUsers Text

The unique ID that Amazon MQ generates for the broker.

Destructuring the Response

data ListUsersResponse

See: newListUsersResponse smart constructor.

Constructors

ListUsersResponse'

Fields

nextToken :: Maybe Text

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

maxResults :: Maybe Natural

Required. The maximum number of ActiveMQ users that can be returned per page (20 by default). This value must be an integer from 5 to 100.

brokerId :: Maybe Text

Required. The unique ID that Amazon MQ generates for the broker.

users :: Maybe [UserSummary]

Required. The list of all ActiveMQ usernames for the specified broker. Does not apply to RabbitMQ brokers.

httpStatus :: Int

The response's http status code.

newListUsersResponse

Arguments

:: Int
$sel:httpStatus:ListUsersResponse'

-> ListUsersResponse

Create a value of ListUsersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListUsers', listUsersResponse_nextToken - The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

$sel:maxResults:ListUsers', listUsersResponse_maxResults - Required. The maximum number of ActiveMQ users that can be returned per page (20 by default). This value must be an integer from 5 to 100.

$sel:brokerId:ListUsers', listUsersResponse_brokerId - Required. The unique ID that Amazon MQ generates for the broker.

$sel:users:ListUsersResponse', listUsersResponse_users - Required. The list of all ActiveMQ usernames for the specified broker. Does not apply to RabbitMQ brokers.

$sel:httpStatus:ListUsersResponse', listUsersResponse_httpStatus - The response's http status code.

Response Lenses

listUsersResponse_nextToken :: Lens' ListUsersResponse (Maybe Text)

The token that specifies the next page of results Amazon MQ should return. To request the first page, leave nextToken empty.

listUsersResponse_maxResults :: Lens' ListUsersResponse (Maybe Natural)

Required. The maximum number of ActiveMQ users that can be returned per page (20 by default). This value must be an integer from 5 to 100.

listUsersResponse_brokerId :: Lens' ListUsersResponse (Maybe Text)

Required. The unique ID that Amazon MQ generates for the broker.

listUsersResponse_users :: Lens' ListUsersResponse (Maybe [UserSummary])

Required. The list of all ActiveMQ usernames for the specified broker. Does not apply to RabbitMQ brokers.

listUsersResponse_httpStatus :: Lens' ListUsersResponse Int

The response's http status code.