Network.AWS.MQ.Types.Configurations

Description

Documentation

data Configurations

Broker configuration information

See: newConfigurations smart constructor.

Constructors

Configurations'

Fields

pending :: Maybe ConfigurationId

The broker's pending configuration.

current :: Maybe ConfigurationId

The broker's current configuration.

history :: Maybe [ConfigurationId]

The history of configurations applied to the broker.

newConfigurations :: Configurations

Create a value of Configurations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pending:Configurations', configurations_pending - The broker's pending configuration.

$sel:current:Configurations', configurations_current - The broker's current configuration.

$sel:history:Configurations', configurations_history - The history of configurations applied to the broker.

configurations_pending :: Lens' Configurations (Maybe ConfigurationId)

The broker's pending configuration.

configurations_current :: Lens' Configurations (Maybe ConfigurationId)

The broker's current configuration.

configurations_history :: Lens' Configurations (Maybe [ConfigurationId])

The history of configurations applied to the broker.