Network.AWS.MQ.Types.LdapServerMetadataInput

Description

Documentation

data LdapServerMetadataInput

Optional. The metadata of the LDAP server used to authenticate and authorize connections to the broker.

Does not apply to RabbitMQ brokers.

See: newLdapServerMetadataInput smart constructor.

Constructors

LdapServerMetadataInput'

Fields

roleName :: Maybe Text

Specifies the LDAP attribute that identifies the group name attribute in the object returned from the group membership query.

userSearchSubtree :: Maybe Bool

The directory search scope for the user. If set to true, scope is to search the entire subtree.

userRoleName :: Maybe Text

Specifies the name of the LDAP attribute for the user group membership.

roleSearchSubtree :: Maybe Bool

The directory search scope for the role. If set to true, scope is to search the entire subtree.

hosts :: [Text]

Specifies the location of the LDAP server such as AWS Directory Service for Microsoft Active Directory . Optional failover server.

userSearchMatching :: Text

The LDAP search filter used to find users within the userBase. The client's username is substituted into the {0} placeholder in the search filter. For example, if this option is set to (uid={0}) and the received username is janedoe, the search filter becomes (uid=janedoe) after string substitution. It will result in matching an entry like uid=janedoe, ou=Users,ou=corp, dc=corp, dc=example, dc=com.

userBase :: Text

Select a particular subtree of the directory information tree (DIT) to search for user entries. The subtree is specified by a DN, which specifies the base node of the subtree. For example, by setting this option to ou=Users,ou=corp, dc=corp, dc=example, dc=com, the search for user entries is restricted to the subtree beneath ou=Users, ou=corp, dc=corp, dc=example, dc=com.

roleSearchMatching :: Text

The LDAP search filter used to find roles within the roleBase. The distinguished name of the user matched by userSearchMatching is substituted into the {0} placeholder in the search filter. The client's username is substituted into the {1} placeholder. For example, if you set this option to (member=uid={1})for the user janedoe, the search filter becomes (member=uid=janedoe) after string substitution. It matches all role entries that have a member attribute equal to uid=janedoe under the subtree selected by the roleBase.

serviceAccountUsername :: Text

Service account username. A service account is an account in your LDAP server that has access to initiate a connection. For example, cn=admin,dc=corp, dc=example, dc=com.

roleBase :: Text

The distinguished name of the node in the directory information tree (DIT) to search for roles or groups. For example, ou=group, ou=corp, dc=corp, dc=example, dc=com.

serviceAccountPassword :: Text

Service account password. A service account is an account in your LDAP server that has access to initiate a connection. For example, cn=admin,dc=corp, dc=example, dc=com.

newLdapServerMetadataInput

Arguments

:: Text
$sel:userSearchMatching:LdapServerMetadataInput'

-> Text
$sel:userBase:LdapServerMetadataInput'

-> Text
$sel:roleSearchMatching:LdapServerMetadataInput'

-> Text
$sel:serviceAccountUsername:LdapServerMetadataInput'

-> Text
$sel:roleBase:LdapServerMetadataInput'

-> Text
$sel:serviceAccountPassword:LdapServerMetadataInput'

-> LdapServerMetadataInput

Create a value of LdapServerMetadataInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleName:LdapServerMetadataInput', ldapServerMetadataInput_roleName - Specifies the LDAP attribute that identifies the group name attribute in the object returned from the group membership query.

$sel:userSearchSubtree:LdapServerMetadataInput', ldapServerMetadataInput_userSearchSubtree - The directory search scope for the user. If set to true, scope is to search the entire subtree.

$sel:userRoleName:LdapServerMetadataInput', ldapServerMetadataInput_userRoleName - Specifies the name of the LDAP attribute for the user group membership.

$sel:roleSearchSubtree:LdapServerMetadataInput', ldapServerMetadataInput_roleSearchSubtree - The directory search scope for the role. If set to true, scope is to search the entire subtree.

$sel:hosts:LdapServerMetadataInput', ldapServerMetadataInput_hosts - Specifies the location of the LDAP server such as AWS Directory Service for Microsoft Active Directory . Optional failover server.

$sel:userSearchMatching:LdapServerMetadataInput', ldapServerMetadataInput_userSearchMatching - The LDAP search filter used to find users within the userBase. The client's username is substituted into the {0} placeholder in the search filter. For example, if this option is set to (uid={0}) and the received username is janedoe, the search filter becomes (uid=janedoe) after string substitution. It will result in matching an entry like uid=janedoe, ou=Users,ou=corp, dc=corp, dc=example, dc=com.

$sel:userBase:LdapServerMetadataInput', ldapServerMetadataInput_userBase - Select a particular subtree of the directory information tree (DIT) to search for user entries. The subtree is specified by a DN, which specifies the base node of the subtree. For example, by setting this option to ou=Users,ou=corp, dc=corp, dc=example, dc=com, the search for user entries is restricted to the subtree beneath ou=Users, ou=corp, dc=corp, dc=example, dc=com.

$sel:roleSearchMatching:LdapServerMetadataInput', ldapServerMetadataInput_roleSearchMatching - The LDAP search filter used to find roles within the roleBase. The distinguished name of the user matched by userSearchMatching is substituted into the {0} placeholder in the search filter. The client's username is substituted into the {1} placeholder. For example, if you set this option to (member=uid={1})for the user janedoe, the search filter becomes (member=uid=janedoe) after string substitution. It matches all role entries that have a member attribute equal to uid=janedoe under the subtree selected by the roleBase.

$sel:serviceAccountUsername:LdapServerMetadataInput', ldapServerMetadataInput_serviceAccountUsername - Service account username. A service account is an account in your LDAP server that has access to initiate a connection. For example, cn=admin,dc=corp, dc=example, dc=com.

$sel:roleBase:LdapServerMetadataInput', ldapServerMetadataInput_roleBase - The distinguished name of the node in the directory information tree (DIT) to search for roles or groups. For example, ou=group, ou=corp, dc=corp, dc=example, dc=com.

$sel:serviceAccountPassword:LdapServerMetadataInput', ldapServerMetadataInput_serviceAccountPassword - Service account password. A service account is an account in your LDAP server that has access to initiate a connection. For example, cn=admin,dc=corp, dc=example, dc=com.

ldapServerMetadataInput_roleName :: Lens' LdapServerMetadataInput (Maybe Text)

Specifies the LDAP attribute that identifies the group name attribute in the object returned from the group membership query.

ldapServerMetadataInput_userSearchSubtree :: Lens' LdapServerMetadataInput (Maybe Bool)

The directory search scope for the user. If set to true, scope is to search the entire subtree.

ldapServerMetadataInput_userRoleName :: Lens' LdapServerMetadataInput (Maybe Text)

Specifies the name of the LDAP attribute for the user group membership.

ldapServerMetadataInput_roleSearchSubtree :: Lens' LdapServerMetadataInput (Maybe Bool)

The directory search scope for the role. If set to true, scope is to search the entire subtree.

ldapServerMetadataInput_hosts :: Lens' LdapServerMetadataInput [Text]

Specifies the location of the LDAP server such as AWS Directory Service for Microsoft Active Directory . Optional failover server.

ldapServerMetadataInput_userSearchMatching :: Lens' LdapServerMetadataInput Text

The LDAP search filter used to find users within the userBase. The client's username is substituted into the {0} placeholder in the search filter. For example, if this option is set to (uid={0}) and the received username is janedoe, the search filter becomes (uid=janedoe) after string substitution. It will result in matching an entry like uid=janedoe, ou=Users,ou=corp, dc=corp, dc=example, dc=com.

ldapServerMetadataInput_userBase :: Lens' LdapServerMetadataInput Text

Select a particular subtree of the directory information tree (DIT) to search for user entries. The subtree is specified by a DN, which specifies the base node of the subtree. For example, by setting this option to ou=Users,ou=corp, dc=corp, dc=example, dc=com, the search for user entries is restricted to the subtree beneath ou=Users, ou=corp, dc=corp, dc=example, dc=com.

ldapServerMetadataInput_roleSearchMatching :: Lens' LdapServerMetadataInput Text

The LDAP search filter used to find roles within the roleBase. The distinguished name of the user matched by userSearchMatching is substituted into the {0} placeholder in the search filter. The client's username is substituted into the {1} placeholder. For example, if you set this option to (member=uid={1})for the user janedoe, the search filter becomes (member=uid=janedoe) after string substitution. It matches all role entries that have a member attribute equal to uid=janedoe under the subtree selected by the roleBase.

ldapServerMetadataInput_serviceAccountUsername :: Lens' LdapServerMetadataInput Text

Service account username. A service account is an account in your LDAP server that has access to initiate a connection. For example, cn=admin,dc=corp, dc=example, dc=com.

ldapServerMetadataInput_roleBase :: Lens' LdapServerMetadataInput Text

The distinguished name of the node in the directory information tree (DIT) to search for roles or groups. For example, ou=group, ou=corp, dc=corp, dc=example, dc=com.

ldapServerMetadataInput_serviceAccountPassword :: Lens' LdapServerMetadataInput Text

Service account password. A service account is an account in your LDAP server that has access to initiate a connection. For example, cn=admin,dc=corp, dc=example, dc=com.