Network.AWS.MQ.Types.LogsSummary

Description

Documentation

data LogsSummary

The list of information about logs currently enabled and pending to be deployed for the specified broker.

See: newLogsSummary smart constructor.

Constructors

LogsSummary'

Fields

audit :: Maybe Bool

Enables audit logging. Every user management action made using JMX or the ActiveMQ Web Console is logged.

pending :: Maybe PendingLogs

The list of information about logs pending to be deployed for the specified broker.

auditLogGroup :: Maybe Text

The location of the CloudWatch Logs log group where audit logs are sent.

generalLogGroup :: Text

The location of the CloudWatch Logs log group where general logs are sent.

general :: Bool

Enables general logging.

newLogsSummary

Arguments

:: Text
$sel:generalLogGroup:LogsSummary'

-> Bool
$sel:general:LogsSummary'

-> LogsSummary

Create a value of LogsSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:audit:LogsSummary', logsSummary_audit - Enables audit logging. Every user management action made using JMX or the ActiveMQ Web Console is logged.

$sel:pending:LogsSummary', logsSummary_pending - The list of information about logs pending to be deployed for the specified broker.

$sel:auditLogGroup:LogsSummary', logsSummary_auditLogGroup - The location of the CloudWatch Logs log group where audit logs are sent.

$sel:generalLogGroup:LogsSummary', logsSummary_generalLogGroup - The location of the CloudWatch Logs log group where general logs are sent.

$sel:general:LogsSummary', logsSummary_general - Enables general logging.

logsSummary_audit :: Lens' LogsSummary (Maybe Bool)

Enables audit logging. Every user management action made using JMX or the ActiveMQ Web Console is logged.

logsSummary_pending :: Lens' LogsSummary (Maybe PendingLogs)

The list of information about logs pending to be deployed for the specified broker.

logsSummary_auditLogGroup :: Lens' LogsSummary (Maybe Text)

The location of the CloudWatch Logs log group where audit logs are sent.

logsSummary_generalLogGroup :: Lens' LogsSummary Text

The location of the CloudWatch Logs log group where general logs are sent.

logsSummary_general :: Lens' LogsSummary Bool

Enables general logging.