Network.AWS.MQ.Types.SanitizationWarning

Description

Documentation

data SanitizationWarning

Returns information about the XML element or attribute that was sanitized in the configuration.

See: newSanitizationWarning smart constructor.

Constructors

SanitizationWarning'

Fields

elementName :: Maybe Text

The name of the XML element that has been sanitized.

attributeName :: Maybe Text

The name of the XML attribute that has been sanitized.

reason :: SanitizationWarningReason

Required. The reason for which the XML elements or attributes were sanitized.

newSanitizationWarning

Arguments

:: SanitizationWarningReason
$sel:reason:SanitizationWarning'

-> SanitizationWarning

Create a value of SanitizationWarning with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:elementName:SanitizationWarning', sanitizationWarning_elementName - The name of the XML element that has been sanitized.

$sel:attributeName:SanitizationWarning', sanitizationWarning_attributeName - The name of the XML attribute that has been sanitized.

$sel:reason:SanitizationWarning', sanitizationWarning_reason - Required. The reason for which the XML elements or attributes were sanitized.

sanitizationWarning_elementName :: Lens' SanitizationWarning (Maybe Text)

The name of the XML element that has been sanitized.

sanitizationWarning_attributeName :: Lens' SanitizationWarning (Maybe Text)

The name of the XML attribute that has been sanitized.

sanitizationWarning_reason :: Lens' SanitizationWarning SanitizationWarningReason

Required. The reason for which the XML elements or attributes were sanitized.