Network.AWS.MQ.UpdateBroker

Description

Adds a pending configuration change to a broker.

Creating a Request

data UpdateBroker

Updates the broker using the specified properties.

See: newUpdateBroker smart constructor.

Constructors

UpdateBroker'

Fields

ldapServerMetadata :: Maybe LdapServerMetadataInput

Optional. The metadata of the LDAP server used to authenticate and authorize connections to the broker. Does not apply to RabbitMQ brokers.

authenticationStrategy :: Maybe AuthenticationStrategy

Optional. The authentication strategy used to secure the broker. The default is SIMPLE.

configuration :: Maybe ConfigurationId

A list of information about the configuration.

securityGroups :: Maybe [Text]

The list of security groups (1 minimum, 5 maximum) that authorizes connections to brokers.

logs :: Maybe Logs

Enables Amazon CloudWatch logging for brokers.

engineVersion :: Maybe Text

The broker engine version. For a list of supported engine versions, see Supported engines.

maintenanceWindowStartTime :: Maybe WeeklyStartTime

The parameters that determine the WeeklyStartTime.

hostInstanceType :: Maybe Text

The broker's host instance type to upgrade to. For a list of supported instance types, see Broker instance types.

autoMinorVersionUpgrade :: Maybe Bool

Enables automatic upgrades to new minor versions for brokers, as new versions are released and supported by Amazon MQ. Automatic upgrades occur during the scheduled maintenance window of the broker or after a manual broker reboot.

brokerId :: Text

The unique ID that Amazon MQ generates for the broker.

newUpdateBroker

Arguments

:: Text
$sel:brokerId:UpdateBroker'

-> UpdateBroker

Create a value of UpdateBroker with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ldapServerMetadata:UpdateBroker', updateBroker_ldapServerMetadata - Optional. The metadata of the LDAP server used to authenticate and authorize connections to the broker. Does not apply to RabbitMQ brokers.

$sel:authenticationStrategy:UpdateBroker', updateBroker_authenticationStrategy - Optional. The authentication strategy used to secure the broker. The default is SIMPLE.

$sel:configuration:UpdateBroker', updateBroker_configuration - A list of information about the configuration.

$sel:securityGroups:UpdateBroker', updateBroker_securityGroups - The list of security groups (1 minimum, 5 maximum) that authorizes connections to brokers.

$sel:logs:UpdateBroker', updateBroker_logs - Enables Amazon CloudWatch logging for brokers.

$sel:engineVersion:UpdateBroker', updateBroker_engineVersion - The broker engine version. For a list of supported engine versions, see Supported engines.

$sel:maintenanceWindowStartTime:UpdateBroker', updateBroker_maintenanceWindowStartTime - The parameters that determine the WeeklyStartTime.

$sel:hostInstanceType:UpdateBroker', updateBroker_hostInstanceType - The broker's host instance type to upgrade to. For a list of supported instance types, see Broker instance types.

$sel:autoMinorVersionUpgrade:UpdateBroker', updateBroker_autoMinorVersionUpgrade - Enables automatic upgrades to new minor versions for brokers, as new versions are released and supported by Amazon MQ. Automatic upgrades occur during the scheduled maintenance window of the broker or after a manual broker reboot.

$sel:brokerId:UpdateBroker', updateBroker_brokerId - The unique ID that Amazon MQ generates for the broker.

Request Lenses

updateBroker_ldapServerMetadata :: Lens' UpdateBroker (Maybe LdapServerMetadataInput)

Optional. The metadata of the LDAP server used to authenticate and authorize connections to the broker. Does not apply to RabbitMQ brokers.

updateBroker_authenticationStrategy :: Lens' UpdateBroker (Maybe AuthenticationStrategy)

Optional. The authentication strategy used to secure the broker. The default is SIMPLE.

updateBroker_configuration :: Lens' UpdateBroker (Maybe ConfigurationId)

A list of information about the configuration.

updateBroker_securityGroups :: Lens' UpdateBroker (Maybe [Text])

The list of security groups (1 minimum, 5 maximum) that authorizes connections to brokers.

updateBroker_logs :: Lens' UpdateBroker (Maybe Logs)

Enables Amazon CloudWatch logging for brokers.

updateBroker_engineVersion :: Lens' UpdateBroker (Maybe Text)

The broker engine version. For a list of supported engine versions, see Supported engines.

updateBroker_maintenanceWindowStartTime :: Lens' UpdateBroker (Maybe WeeklyStartTime)

The parameters that determine the WeeklyStartTime.

updateBroker_hostInstanceType :: Lens' UpdateBroker (Maybe Text)

The broker's host instance type to upgrade to. For a list of supported instance types, see Broker instance types.

updateBroker_autoMinorVersionUpgrade :: Lens' UpdateBroker (Maybe Bool)

Enables automatic upgrades to new minor versions for brokers, as new versions are released and supported by Amazon MQ. Automatic upgrades occur during the scheduled maintenance window of the broker or after a manual broker reboot.

updateBroker_brokerId :: Lens' UpdateBroker Text

The unique ID that Amazon MQ generates for the broker.

Destructuring the Response

data UpdateBrokerResponse

See: newUpdateBrokerResponse smart constructor.

Constructors

UpdateBrokerResponse'

Fields

ldapServerMetadata :: Maybe LdapServerMetadataOutput

Optional. The metadata of the LDAP server used to authenticate and authorize connections to the broker. Does not apply to RabbitMQ brokers.

brokerId :: Maybe Text

Required. The unique ID that Amazon MQ generates for the broker.

authenticationStrategy :: Maybe AuthenticationStrategy

Optional. The authentication strategy used to secure the broker. The default is SIMPLE.

configuration :: Maybe ConfigurationId

The ID of the updated configuration.

securityGroups :: Maybe [Text]

The list of security groups (1 minimum, 5 maximum) that authorizes connections to brokers.

logs :: Maybe Logs

The list of information about logs to be enabled for the specified broker.

engineVersion :: Maybe Text

The broker engine version to upgrade to. For a list of supported engine versions, see Supported engines.

maintenanceWindowStartTime :: Maybe WeeklyStartTime

The parameters that determine the WeeklyStartTime.

hostInstanceType :: Maybe Text

The broker's host instance type to upgrade to. For a list of supported instance types, see Broker instance types.

autoMinorVersionUpgrade :: Maybe Bool

The new boolean value that specifies whether broker engines automatically upgrade to new minor versions as new versions are released and supported by Amazon MQ.

httpStatus :: Int

The response's http status code.

newUpdateBrokerResponse

Arguments

:: Int
$sel:httpStatus:UpdateBrokerResponse'

-> UpdateBrokerResponse

Create a value of UpdateBrokerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ldapServerMetadata:UpdateBroker', updateBrokerResponse_ldapServerMetadata - Optional. The metadata of the LDAP server used to authenticate and authorize connections to the broker. Does not apply to RabbitMQ brokers.

$sel:brokerId:UpdateBroker', updateBrokerResponse_brokerId - Required. The unique ID that Amazon MQ generates for the broker.

$sel:authenticationStrategy:UpdateBroker', updateBrokerResponse_authenticationStrategy - Optional. The authentication strategy used to secure the broker. The default is SIMPLE.

$sel:configuration:UpdateBroker', updateBrokerResponse_configuration - The ID of the updated configuration.

$sel:securityGroups:UpdateBroker', updateBrokerResponse_securityGroups - The list of security groups (1 minimum, 5 maximum) that authorizes connections to brokers.

$sel:logs:UpdateBroker', updateBrokerResponse_logs - The list of information about logs to be enabled for the specified broker.

$sel:engineVersion:UpdateBroker', updateBrokerResponse_engineVersion - The broker engine version to upgrade to. For a list of supported engine versions, see Supported engines.

$sel:maintenanceWindowStartTime:UpdateBroker', updateBrokerResponse_maintenanceWindowStartTime - The parameters that determine the WeeklyStartTime.

$sel:hostInstanceType:UpdateBroker', updateBrokerResponse_hostInstanceType - The broker's host instance type to upgrade to. For a list of supported instance types, see Broker instance types.

$sel:autoMinorVersionUpgrade:UpdateBroker', updateBrokerResponse_autoMinorVersionUpgrade - The new boolean value that specifies whether broker engines automatically upgrade to new minor versions as new versions are released and supported by Amazon MQ.

$sel:httpStatus:UpdateBrokerResponse', updateBrokerResponse_httpStatus - The response's http status code.

Response Lenses

updateBrokerResponse_ldapServerMetadata :: Lens' UpdateBrokerResponse (Maybe LdapServerMetadataOutput)

Optional. The metadata of the LDAP server used to authenticate and authorize connections to the broker. Does not apply to RabbitMQ brokers.

updateBrokerResponse_brokerId :: Lens' UpdateBrokerResponse (Maybe Text)

Required. The unique ID that Amazon MQ generates for the broker.

updateBrokerResponse_authenticationStrategy :: Lens' UpdateBrokerResponse (Maybe AuthenticationStrategy)

Optional. The authentication strategy used to secure the broker. The default is SIMPLE.

updateBrokerResponse_configuration :: Lens' UpdateBrokerResponse (Maybe ConfigurationId)

The ID of the updated configuration.

updateBrokerResponse_securityGroups :: Lens' UpdateBrokerResponse (Maybe [Text])

The list of security groups (1 minimum, 5 maximum) that authorizes connections to brokers.

updateBrokerResponse_logs :: Lens' UpdateBrokerResponse (Maybe Logs)

The list of information about logs to be enabled for the specified broker.

updateBrokerResponse_engineVersion :: Lens' UpdateBrokerResponse (Maybe Text)

The broker engine version to upgrade to. For a list of supported engine versions, see Supported engines.

updateBrokerResponse_maintenanceWindowStartTime :: Lens' UpdateBrokerResponse (Maybe WeeklyStartTime)

The parameters that determine the WeeklyStartTime.

updateBrokerResponse_hostInstanceType :: Lens' UpdateBrokerResponse (Maybe Text)

The broker's host instance type to upgrade to. For a list of supported instance types, see Broker instance types.

updateBrokerResponse_autoMinorVersionUpgrade :: Lens' UpdateBrokerResponse (Maybe Bool)

The new boolean value that specifies whether broker engines automatically upgrade to new minor versions as new versions are released and supported by Amazon MQ.

updateBrokerResponse_httpStatus :: Lens' UpdateBrokerResponse Int

The response's http status code.