Network.AWS.MachineLearning.CreateDataSourceFromS

Description

Creates a DataSource object. A DataSource references data that can be used to perform CreateMLModel, CreateEvaluation, or CreateBatchPrediction operations.

CreateDataSourceFromS3 is an asynchronous operation. In response to CreateDataSourceFromS3, Amazon Machine Learning (Amazon ML) immediately returns and sets the DataSource status to PENDING. After the DataSource has been created and is ready for use, Amazon ML sets the Status parameter to COMPLETED. DataSource in the COMPLETED or PENDING state can be used to perform only CreateMLModel, CreateEvaluation or CreateBatchPrediction operations.

If Amazon ML can't accept the input source, it sets the Status parameter to FAILED and includes an error message in the Message attribute of the GetDataSource operation response.

The observation data used in a DataSource should be ready to use; that is, it should have a consistent structure, and missing data values should be kept to a minimum. The observation data must reside in one or more .csv files in an Amazon Simple Storage Service (Amazon S3) location, along with a schema that describes the data items by name and type. The same schema must be used for all of the data files referenced by the DataSource.

After the DataSource has been created, it's ready to use in evaluations and batch predictions. If you plan to use the DataSource to train an MLModel, the DataSource also needs a recipe. A recipe describes how each input variable will be used in training an MLModel. Will the variable be included or excluded from training? Will the variable be manipulated; for example, will it be combined with another variable or will it be split apart into word combinations? The recipe provides answers to these questions.

Creating a Request

data CreateDataSourceFromS

See: newCreateDataSourceFromS smart constructor.

Constructors

CreateDataSourceFromS'

Fields

computeStatistics :: Maybe Bool

The compute statistics for a DataSource. The statistics are generated from the observation data referenced by a DataSource. Amazon ML uses the statistics internally during MLModel training. This parameter must be set to true if the @DataSource needs to be used for MLModel@ training.

dataSourceName :: Maybe Text

A user-supplied name or description of the DataSource.

dataSourceId :: Text

A user-supplied identifier that uniquely identifies the DataSource.

dataSpec :: S3DataSpec

The data specification of a DataSource:

DataLocationS3 - The Amazon S3 location of the observation data.

DataSchemaLocationS3 - The Amazon S3 location of the DataSchema.

DataSchema - A JSON string representing the schema. This is not required if DataSchemaUri is specified.

DataRearrangement - A JSON string that represents the splitting and rearrangement requirements for the Datasource.

Sample - "{\"splitting\":{\"percentBegin\":10,\"percentEnd\":60}}"

newCreateDataSourceFromS

Arguments

:: Text
$sel:dataSourceId:CreateDataSourceFromS'

-> S3DataSpec
$sel:dataSpec:CreateDataSourceFromS'

-> CreateDataSourceFromS

Create a value of CreateDataSourceFromS with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:computeStatistics:CreateDataSourceFromS', createDataSourceFromS_computeStatistics - The compute statistics for a DataSource. The statistics are generated from the observation data referenced by a DataSource. Amazon ML uses the statistics internally during MLModel training. This parameter must be set to true if the @DataSource needs to be used for MLModel@ training.

$sel:dataSourceName:CreateDataSourceFromS', createDataSourceFromS_dataSourceName - A user-supplied name or description of the DataSource.

$sel:dataSourceId:CreateDataSourceFromS', createDataSourceFromS_dataSourceId - A user-supplied identifier that uniquely identifies the DataSource.

$sel:dataSpec:CreateDataSourceFromS', createDataSourceFromS_dataSpec - The data specification of a DataSource:

DataLocationS3 - The Amazon S3 location of the observation data.

DataSchemaLocationS3 - The Amazon S3 location of the DataSchema.

DataSchema - A JSON string representing the schema. This is not required if DataSchemaUri is specified.

DataRearrangement - A JSON string that represents the splitting and rearrangement requirements for the Datasource.

Sample - "{\"splitting\":{\"percentBegin\":10,\"percentEnd\":60}}"

Request Lenses

createDataSourceFromS_computeStatistics :: Lens' CreateDataSourceFromS (Maybe Bool)

The compute statistics for a DataSource. The statistics are generated from the observation data referenced by a DataSource. Amazon ML uses the statistics internally during MLModel training. This parameter must be set to true if the @DataSource needs to be used for MLModel@ training.

createDataSourceFromS_dataSourceName :: Lens' CreateDataSourceFromS (Maybe Text)

A user-supplied name or description of the DataSource.

createDataSourceFromS_dataSourceId :: Lens' CreateDataSourceFromS Text

A user-supplied identifier that uniquely identifies the DataSource.

createDataSourceFromS_dataSpec :: Lens' CreateDataSourceFromS S3DataSpec

The data specification of a DataSource:

DataLocationS3 - The Amazon S3 location of the observation data.

DataSchemaLocationS3 - The Amazon S3 location of the DataSchema.

DataSchema - A JSON string representing the schema. This is not required if DataSchemaUri is specified.

DataRearrangement - A JSON string that represents the splitting and rearrangement requirements for the Datasource.

Sample - "{\"splitting\":{\"percentBegin\":10,\"percentEnd\":60}}"

Destructuring the Response

data CreateDataSourceFromSResponse

Represents the output of a CreateDataSourceFromS3 operation, and is an acknowledgement that Amazon ML received the request.

The CreateDataSourceFromS3 operation is asynchronous. You can poll for updates by using the GetBatchPrediction operation and checking the Status parameter.

See: newCreateDataSourceFromSResponse smart constructor.

Constructors

CreateDataSourceFromSResponse'

Fields

dataSourceId :: Maybe Text

A user-supplied ID that uniquely identifies the DataSource. This value should be identical to the value of the DataSourceID in the request.

httpStatus :: Int

The response's http status code.

newCreateDataSourceFromSResponse

Arguments

:: Int
$sel:httpStatus:CreateDataSourceFromSResponse'

-> CreateDataSourceFromSResponse

Create a value of CreateDataSourceFromSResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataSourceId:CreateDataSourceFromS', createDataSourceFromSResponse_dataSourceId - A user-supplied ID that uniquely identifies the DataSource. This value should be identical to the value of the DataSourceID in the request.

$sel:httpStatus:CreateDataSourceFromSResponse', createDataSourceFromSResponse_httpStatus - The response's http status code.

Response Lenses

createDataSourceFromSResponse_dataSourceId :: Lens' CreateDataSourceFromSResponse (Maybe Text)

A user-supplied ID that uniquely identifies the DataSource. This value should be identical to the value of the DataSourceID in the request.

createDataSourceFromSResponse_httpStatus :: Lens' CreateDataSourceFromSResponse Int

The response's http status code.