Network.AWS.MachineLearning.CreateDataSourceFromS3

Description

Creates a DataSource object. A DataSource references data that can be used to perform CreateMLModel, CreateEvaluation, or CreateBatchPrediction operations.

CreateDataSourceFromS3 is an asynchronous operation. In response to CreateDataSourceFromS3, Amazon Machine Learning (Amazon ML) immediately returns and sets the DataSource status to PENDING. After the DataSource has been created and is ready for use, Amazon ML sets the Status parameter to COMPLETED. DataSource in the COMPLETED or PENDING state can be used to perform only CreateMLModel, CreateEvaluation or CreateBatchPrediction operations.

If Amazon ML can't accept the input source, it sets the Status parameter to FAILED and includes an error message in the Message attribute of the GetDataSource operation response.

The observation data used in a DataSource should be ready to use; that is, it should have a consistent structure, and missing data values should be kept to a minimum. The observation data must reside in one or more .csv files in an Amazon Simple Storage Service (Amazon S3) location, along with a schema that describes the data items by name and type. The same schema must be used for all of the data files referenced by the DataSource.

After the DataSource has been created, it's ready to use in evaluations and batch predictions. If you plan to use the DataSource to train an MLModel, the DataSource also needs a recipe. A recipe describes how each input variable will be used in training an MLModel. Will the variable be included or excluded from training? Will the variable be manipulated; for example, will it be combined with another variable or will it be split apart into word combinations? The recipe provides answers to these questions.

Creating a Request

data CreateDataSourceFromS3

See: newCreateDataSourceFromS3 smart constructor.

Constructors

CreateDataSourceFromS3'

Fields

dataSourceName :: Maybe Text

A user-supplied name or description of the DataSource.

computeStatistics :: Maybe Bool

The compute statistics for a DataSource. The statistics are generated from the observation data referenced by a DataSource. Amazon ML uses the statistics internally during MLModel training. This parameter must be set to true if the @DataSource needs to be used for MLModel@ training.

dataSourceId :: Text

A user-supplied identifier that uniquely identifies the DataSource.

dataSpec :: S3DataSpec

The data specification of a DataSource:

DataLocationS3 - The Amazon S3 location of the observation data.

DataSchemaLocationS3 - The Amazon S3 location of the DataSchema.

DataSchema - A JSON string representing the schema. This is not required if DataSchemaUri is specified.

DataRearrangement - A JSON string that represents the splitting and rearrangement requirements for the Datasource.

Sample - "{\"splitting\":{\"percentBegin\":10,\"percentEnd\":60}}"

newCreateDataSourceFromS3

Arguments

:: Text
$sel:dataSourceId:CreateDataSourceFromS3'

-> S3DataSpec
$sel:dataSpec:CreateDataSourceFromS3'

-> CreateDataSourceFromS3

Create a value of CreateDataSourceFromS3 with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataSourceName:CreateDataSourceFromS3', createDataSourceFromS3_dataSourceName - A user-supplied name or description of the DataSource.

$sel:computeStatistics:CreateDataSourceFromS3', createDataSourceFromS3_computeStatistics - The compute statistics for a DataSource. The statistics are generated from the observation data referenced by a DataSource. Amazon ML uses the statistics internally during MLModel training. This parameter must be set to true if the @DataSource needs to be used for MLModel@ training.

$sel:dataSourceId:CreateDataSourceFromS3', createDataSourceFromS3_dataSourceId - A user-supplied identifier that uniquely identifies the DataSource.

$sel:dataSpec:CreateDataSourceFromS3', createDataSourceFromS3_dataSpec - The data specification of a DataSource:

DataLocationS3 - The Amazon S3 location of the observation data.

DataSchemaLocationS3 - The Amazon S3 location of the DataSchema.

DataSchema - A JSON string representing the schema. This is not required if DataSchemaUri is specified.

DataRearrangement - A JSON string that represents the splitting and rearrangement requirements for the Datasource.

Sample - "{\"splitting\":{\"percentBegin\":10,\"percentEnd\":60}}"

Request Lenses

createDataSourceFromS3_dataSourceName :: Lens' CreateDataSourceFromS3 (Maybe Text)

A user-supplied name or description of the DataSource.

createDataSourceFromS3_computeStatistics :: Lens' CreateDataSourceFromS3 (Maybe Bool)

The compute statistics for a DataSource. The statistics are generated from the observation data referenced by a DataSource. Amazon ML uses the statistics internally during MLModel training. This parameter must be set to true if the @DataSource needs to be used for MLModel@ training.

createDataSourceFromS3_dataSourceId :: Lens' CreateDataSourceFromS3 Text

A user-supplied identifier that uniquely identifies the DataSource.

createDataSourceFromS3_dataSpec :: Lens' CreateDataSourceFromS3 S3DataSpec

The data specification of a DataSource:

DataLocationS3 - The Amazon S3 location of the observation data.

DataSchemaLocationS3 - The Amazon S3 location of the DataSchema.

DataSchema - A JSON string representing the schema. This is not required if DataSchemaUri is specified.

DataRearrangement - A JSON string that represents the splitting and rearrangement requirements for the Datasource.

Sample - "{\"splitting\":{\"percentBegin\":10,\"percentEnd\":60}}"

Destructuring the Response

data CreateDataSourceFromS3Response

Represents the output of a CreateDataSourceFromS3 operation, and is an acknowledgement that Amazon ML received the request.

The CreateDataSourceFromS3 operation is asynchronous. You can poll for updates by using the GetBatchPrediction operation and checking the Status parameter.

See: newCreateDataSourceFromS3Response smart constructor.

Constructors

CreateDataSourceFromS3Response'

Fields

dataSourceId :: Maybe Text

A user-supplied ID that uniquely identifies the DataSource. This value should be identical to the value of the DataSourceID in the request.

httpStatus :: Int

The response's http status code.

newCreateDataSourceFromS3Response

Arguments

:: Int
$sel:httpStatus:CreateDataSourceFromS3Response'

-> CreateDataSourceFromS3Response

Create a value of CreateDataSourceFromS3Response with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataSourceId:CreateDataSourceFromS3', createDataSourceFromS3Response_dataSourceId - A user-supplied ID that uniquely identifies the DataSource. This value should be identical to the value of the DataSourceID in the request.

$sel:httpStatus:CreateDataSourceFromS3Response', createDataSourceFromS3Response_httpStatus - The response's http status code.

Response Lenses

createDataSourceFromS3Response_dataSourceId :: Lens' CreateDataSourceFromS3Response (Maybe Text)

A user-supplied ID that uniquely identifies the DataSource. This value should be identical to the value of the DataSourceID in the request.

createDataSourceFromS3Response_httpStatus :: Lens' CreateDataSourceFromS3Response Int

The response's http status code.