Network.AWS.MachineLearning.CreateMLModel

Description

Creates a new MLModel using the DataSource and the recipe as information sources.

An MLModel is nearly immutable. Users can update only the MLModelName and the ScoreThreshold in an MLModel without creating a new MLModel.

CreateMLModel is an asynchronous operation. In response to CreateMLModel, Amazon Machine Learning (Amazon ML) immediately returns and sets the MLModel status to PENDING. After the MLModel has been created and ready is for use, Amazon ML sets the status to COMPLETED.

You can use the GetMLModel operation to check the progress of the MLModel during the creation operation.

CreateMLModel requires a DataSource with computed statistics, which can be created by setting ComputeStatistics to true in CreateDataSourceFromRDS, CreateDataSourceFromS3, or CreateDataSourceFromRedshift operations.

Creating a Request

data CreateMLModel

See: newCreateMLModel smart constructor.

Constructors

CreateMLModel'

Fields

recipeUri :: Maybe Text

The Amazon Simple Storage Service (Amazon S3) location and file name that contains the MLModel recipe. You must specify either the recipe or its URI. If you don't specify a recipe or its URI, Amazon ML creates a default.

recipe :: Maybe Text

The data recipe for creating the MLModel. You must specify either the recipe or its URI. If you don't specify a recipe or its URI, Amazon ML creates a default.

mLModelName :: Maybe Text

A user-supplied name or description of the MLModel.

parameters :: Maybe (HashMap Text Text)

A list of the training parameters in the MLModel. The list is implemented as a map of key-value pairs.

The following is the current set of training parameters:

sgd.maxMLModelSizeInBytes - The maximum allowed size of the model. Depending on the input data, the size of the model might affect its performance.

The value is an integer that ranges from 100000 to 2147483648. The default value is 33554432.

sgd.maxPasses - The number of times that the training process traverses the observations to build the MLModel. The value is an integer that ranges from 1 to 10000. The default value is 10.

sgd.shuffleType - Whether Amazon ML shuffles the training data. Shuffling the data improves a model's ability to find the optimal solution for a variety of data types. The valid values are auto and none. The default value is none. We strongly recommend that you shuffle your data.

sgd.l1RegularizationAmount - The coefficient regularization L1 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to zero, resulting in a sparse feature set. If you use this parameter, start by specifying a small value, such as 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is to not use L1 normalization. This parameter can't be used when L2 is specified. Use this parameter sparingly.

sgd.l2RegularizationAmount - The coefficient regularization L2 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to small, nonzero values. If you use this parameter, start by specifying a small value, such as 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is to not use L2 normalization. This parameter can't be used when L1 is specified. Use this parameter sparingly.

mLModelId :: Text

A user-supplied ID that uniquely identifies the MLModel.

mLModelType :: MLModelType

The category of supervised learning that this MLModel will address. Choose from the following types:

Choose REGRESSION if the MLModel will be used to predict a numeric value.

Choose BINARY if the MLModel result has two possible values.

Choose MULTICLASS if the MLModel result has a limited number of values.

For more information, see the Amazon Machine Learning Developer Guide.

trainingDataSourceId :: Text

The DataSource that points to the training data.

newCreateMLModel

Arguments

:: Text
$sel:mLModelId:CreateMLModel'

-> MLModelType
$sel:mLModelType:CreateMLModel'

-> Text
$sel:trainingDataSourceId:CreateMLModel'

-> CreateMLModel

Create a value of CreateMLModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recipeUri:CreateMLModel', createMLModel_recipeUri - The Amazon Simple Storage Service (Amazon S3) location and file name that contains the MLModel recipe. You must specify either the recipe or its URI. If you don't specify a recipe or its URI, Amazon ML creates a default.

$sel:recipe:CreateMLModel', createMLModel_recipe - The data recipe for creating the MLModel. You must specify either the recipe or its URI. If you don't specify a recipe or its URI, Amazon ML creates a default.

$sel:mLModelName:CreateMLModel', createMLModel_mLModelName - A user-supplied name or description of the MLModel.

$sel:parameters:CreateMLModel', createMLModel_parameters - A list of the training parameters in the MLModel. The list is implemented as a map of key-value pairs.

The following is the current set of training parameters:

sgd.maxMLModelSizeInBytes - The maximum allowed size of the model. Depending on the input data, the size of the model might affect its performance.

The value is an integer that ranges from 100000 to 2147483648. The default value is 33554432.

sgd.maxPasses - The number of times that the training process traverses the observations to build the MLModel. The value is an integer that ranges from 1 to 10000. The default value is 10.

sgd.shuffleType - Whether Amazon ML shuffles the training data. Shuffling the data improves a model's ability to find the optimal solution for a variety of data types. The valid values are auto and none. The default value is none. We strongly recommend that you shuffle your data.

sgd.l1RegularizationAmount - The coefficient regularization L1 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to zero, resulting in a sparse feature set. If you use this parameter, start by specifying a small value, such as 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is to not use L1 normalization. This parameter can't be used when L2 is specified. Use this parameter sparingly.

sgd.l2RegularizationAmount - The coefficient regularization L2 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to small, nonzero values. If you use this parameter, start by specifying a small value, such as 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is to not use L2 normalization. This parameter can't be used when L1 is specified. Use this parameter sparingly.

$sel:mLModelId:CreateMLModel', createMLModel_mLModelId - A user-supplied ID that uniquely identifies the MLModel.

$sel:mLModelType:CreateMLModel', createMLModel_mLModelType - The category of supervised learning that this MLModel will address. Choose from the following types:

Choose REGRESSION if the MLModel will be used to predict a numeric value.

Choose BINARY if the MLModel result has two possible values.

Choose MULTICLASS if the MLModel result has a limited number of values.

For more information, see the Amazon Machine Learning Developer Guide.

$sel:trainingDataSourceId:CreateMLModel', createMLModel_trainingDataSourceId - The DataSource that points to the training data.

Request Lenses

createMLModel_recipeUri :: Lens' CreateMLModel (Maybe Text)

The Amazon Simple Storage Service (Amazon S3) location and file name that contains the MLModel recipe. You must specify either the recipe or its URI. If you don't specify a recipe or its URI, Amazon ML creates a default.

createMLModel_recipe :: Lens' CreateMLModel (Maybe Text)

The data recipe for creating the MLModel. You must specify either the recipe or its URI. If you don't specify a recipe or its URI, Amazon ML creates a default.

createMLModel_mLModelName :: Lens' CreateMLModel (Maybe Text)

A user-supplied name or description of the MLModel.

createMLModel_parameters :: Lens' CreateMLModel (Maybe (HashMap Text Text))

A list of the training parameters in the MLModel. The list is implemented as a map of key-value pairs.

The following is the current set of training parameters:

sgd.maxMLModelSizeInBytes - The maximum allowed size of the model. Depending on the input data, the size of the model might affect its performance.

The value is an integer that ranges from 100000 to 2147483648. The default value is 33554432.

sgd.maxPasses - The number of times that the training process traverses the observations to build the MLModel. The value is an integer that ranges from 1 to 10000. The default value is 10.

sgd.shuffleType - Whether Amazon ML shuffles the training data. Shuffling the data improves a model's ability to find the optimal solution for a variety of data types. The valid values are auto and none. The default value is none. We strongly recommend that you shuffle your data.

sgd.l1RegularizationAmount - The coefficient regularization L1 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to zero, resulting in a sparse feature set. If you use this parameter, start by specifying a small value, such as 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is to not use L1 normalization. This parameter can't be used when L2 is specified. Use this parameter sparingly.

sgd.l2RegularizationAmount - The coefficient regularization L2 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to small, nonzero values. If you use this parameter, start by specifying a small value, such as 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is to not use L2 normalization. This parameter can't be used when L1 is specified. Use this parameter sparingly.

createMLModel_mLModelId :: Lens' CreateMLModel Text

A user-supplied ID that uniquely identifies the MLModel.

createMLModel_mLModelType :: Lens' CreateMLModel MLModelType

The category of supervised learning that this MLModel will address. Choose from the following types:

Choose REGRESSION if the MLModel will be used to predict a numeric value.

Choose BINARY if the MLModel result has two possible values.

Choose MULTICLASS if the MLModel result has a limited number of values.

For more information, see the Amazon Machine Learning Developer Guide.

createMLModel_trainingDataSourceId :: Lens' CreateMLModel Text

The DataSource that points to the training data.

Destructuring the Response

data CreateMLModelResponse

Represents the output of a CreateMLModel operation, and is an acknowledgement that Amazon ML received the request.

The CreateMLModel operation is asynchronous. You can poll for status updates by using the GetMLModel operation and checking the Status parameter.

See: newCreateMLModelResponse smart constructor.

Constructors

CreateMLModelResponse'

Fields

mLModelId :: Maybe Text

A user-supplied ID that uniquely identifies the MLModel. This value should be identical to the value of the MLModelId in the request.

httpStatus :: Int

The response's http status code.

newCreateMLModelResponse

Arguments

:: Int
$sel:httpStatus:CreateMLModelResponse'

-> CreateMLModelResponse

Create a value of CreateMLModelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mLModelId:CreateMLModel', createMLModelResponse_mLModelId - A user-supplied ID that uniquely identifies the MLModel. This value should be identical to the value of the MLModelId in the request.

$sel:httpStatus:CreateMLModelResponse', createMLModelResponse_httpStatus - The response's http status code.

Response Lenses

createMLModelResponse_mLModelId :: Lens' CreateMLModelResponse (Maybe Text)

A user-supplied ID that uniquely identifies the MLModel. This value should be identical to the value of the MLModelId in the request.

createMLModelResponse_httpStatus :: Lens' CreateMLModelResponse Int

The response's http status code.