Network.AWS.MachineLearning.DeleteEvaluation

Description

Assigns the DELETED status to an Evaluation, rendering it unusable.

After invoking the DeleteEvaluation operation, you can use the GetEvaluation operation to verify that the status of the Evaluation changed to DELETED.

Caution: The results of the DeleteEvaluation operation are irreversible.

Creating a Request

data DeleteEvaluation

See: newDeleteEvaluation smart constructor.

Constructors

DeleteEvaluation'

Fields

evaluationId :: Text

A user-supplied ID that uniquely identifies the Evaluation to delete.

newDeleteEvaluation

Arguments

:: Text
$sel:evaluationId:DeleteEvaluation'

-> DeleteEvaluation

Create a value of DeleteEvaluation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:evaluationId:DeleteEvaluation', deleteEvaluation_evaluationId - A user-supplied ID that uniquely identifies the Evaluation to delete.

Request Lenses

deleteEvaluation_evaluationId :: Lens' DeleteEvaluation Text

A user-supplied ID that uniquely identifies the Evaluation to delete.

Destructuring the Response

data DeleteEvaluationResponse

Represents the output of a DeleteEvaluation operation. The output indicates that Amazon Machine Learning (Amazon ML) received the request.

You can use the GetEvaluation operation and check the value of the Status parameter to see whether an Evaluation is marked as DELETED.

See: newDeleteEvaluationResponse smart constructor.

Constructors

DeleteEvaluationResponse'

Fields

evaluationId :: Maybe Text

A user-supplied ID that uniquely identifies the Evaluation. This value should be identical to the value of the EvaluationId in the request.

httpStatus :: Int

The response's http status code.

newDeleteEvaluationResponse

Arguments

:: Int
$sel:httpStatus:DeleteEvaluationResponse'

-> DeleteEvaluationResponse

Create a value of DeleteEvaluationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:evaluationId:DeleteEvaluation', deleteEvaluationResponse_evaluationId - A user-supplied ID that uniquely identifies the Evaluation. This value should be identical to the value of the EvaluationId in the request.

$sel:httpStatus:DeleteEvaluationResponse', deleteEvaluationResponse_httpStatus - The response's http status code.

Response Lenses

deleteEvaluationResponse_evaluationId :: Lens' DeleteEvaluationResponse (Maybe Text)

A user-supplied ID that uniquely identifies the Evaluation. This value should be identical to the value of the EvaluationId in the request.

deleteEvaluationResponse_httpStatus :: Lens' DeleteEvaluationResponse Int

The response's http status code.