Network.AWS.MachineLearning.DescribeBatchPredictions

Description

Returns a list of BatchPrediction operations that match the search criteria in the request.

This operation returns paginated results.

Creating a Request

data DescribeBatchPredictions

See: newDescribeBatchPredictions smart constructor.

Constructors

DescribeBatchPredictions'

Fields

eq :: Maybe Text

The equal to operator. The BatchPrediction results will have FilterVariable values that exactly match the value specified with EQ.

nextToken :: Maybe Text

An ID of the page in the paginated results.

sortOrder :: Maybe SortOrder

A two-value parameter that determines the sequence of the resulting list of MLModels.

asc - Arranges the list in ascending order (A-Z, 0-9).

dsc - Arranges the list in descending order (Z-A, 9-0).

Results are sorted by FilterVariable.

filterVariable :: Maybe BatchPredictionFilterVariable

Use one of the following variables to filter a list of BatchPrediction:

CreatedAt - Sets the search criteria to the BatchPrediction creation date.

Status - Sets the search criteria to the BatchPrediction status.

Name - Sets the search criteria to the contents of the BatchPrediction ____ Name.

IAMUser - Sets the search criteria to the user account that invoked the BatchPrediction creation.

MLModelId - Sets the search criteria to the MLModel used in the BatchPrediction.

DataSourceId - Sets the search criteria to the DataSource used in the BatchPrediction.

DataURI - Sets the search criteria to the data file(s) used in the BatchPrediction. The URL can identify either a file or an Amazon Simple Storage Solution (Amazon S3) bucket or directory.

gt :: Maybe Text

The greater than operator. The BatchPrediction results will have FilterVariable values that are greater than the value specified with GT.

ne :: Maybe Text

The not equal to operator. The BatchPrediction results will have FilterVariable values not equal to the value specified with NE.

prefix :: Maybe Text

A string that is found at the beginning of a variable, such as Name or Id.

For example, a Batch Prediction operation could have the Name 2014-09-09-HolidayGiftMailer. To search for this BatchPrediction, select Name for the FilterVariable and any of the following strings for the Prefix:

2014-09

2014-09-09

2014-09-09-Holiday

ge :: Maybe Text

The greater than or equal to operator. The BatchPrediction results will have FilterVariable values that are greater than or equal to the value specified with GE.

le :: Maybe Text

The less than or equal to operator. The BatchPrediction results will have FilterVariable values that are less than or equal to the value specified with LE.

lt :: Maybe Text

The less than operator. The BatchPrediction results will have FilterVariable values that are less than the value specified with LT.

limit :: Maybe Natural

The number of pages of information to include in the result. The range of acceptable values is 1 through 100. The default value is 100.

newDescribeBatchPredictions :: DescribeBatchPredictions

Create a value of DescribeBatchPredictions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eq:DescribeBatchPredictions', describeBatchPredictions_eq - The equal to operator. The BatchPrediction results will have FilterVariable values that exactly match the value specified with EQ.

$sel:nextToken:DescribeBatchPredictions', describeBatchPredictions_nextToken - An ID of the page in the paginated results.

$sel:sortOrder:DescribeBatchPredictions', describeBatchPredictions_sortOrder - A two-value parameter that determines the sequence of the resulting list of MLModels.

asc - Arranges the list in ascending order (A-Z, 0-9).

dsc - Arranges the list in descending order (Z-A, 9-0).

Results are sorted by FilterVariable.

$sel:filterVariable:DescribeBatchPredictions', describeBatchPredictions_filterVariable - Use one of the following variables to filter a list of BatchPrediction:

CreatedAt - Sets the search criteria to the BatchPrediction creation date.

Status - Sets the search criteria to the BatchPrediction status.

Name - Sets the search criteria to the contents of the BatchPrediction ____ Name.

IAMUser - Sets the search criteria to the user account that invoked the BatchPrediction creation.

MLModelId - Sets the search criteria to the MLModel used in the BatchPrediction.

DataSourceId - Sets the search criteria to the DataSource used in the BatchPrediction.

DataURI - Sets the search criteria to the data file(s) used in the BatchPrediction. The URL can identify either a file or an Amazon Simple Storage Solution (Amazon S3) bucket or directory.

$sel:gt:DescribeBatchPredictions', describeBatchPredictions_gt - The greater than operator. The BatchPrediction results will have FilterVariable values that are greater than the value specified with GT.

$sel:ne:DescribeBatchPredictions', describeBatchPredictions_ne - The not equal to operator. The BatchPrediction results will have FilterVariable values not equal to the value specified with NE.

$sel:prefix:DescribeBatchPredictions', describeBatchPredictions_prefix - A string that is found at the beginning of a variable, such as Name or Id.

For example, a Batch Prediction operation could have the Name 2014-09-09-HolidayGiftMailer. To search for this BatchPrediction, select Name for the FilterVariable and any of the following strings for the Prefix:

2014-09

2014-09-09

2014-09-09-Holiday

$sel:ge:DescribeBatchPredictions', describeBatchPredictions_ge - The greater than or equal to operator. The BatchPrediction results will have FilterVariable values that are greater than or equal to the value specified with GE.

$sel:le:DescribeBatchPredictions', describeBatchPredictions_le - The less than or equal to operator. The BatchPrediction results will have FilterVariable values that are less than or equal to the value specified with LE.

$sel:lt:DescribeBatchPredictions', describeBatchPredictions_lt - The less than operator. The BatchPrediction results will have FilterVariable values that are less than the value specified with LT.

$sel:limit:DescribeBatchPredictions', describeBatchPredictions_limit - The number of pages of information to include in the result. The range of acceptable values is 1 through 100. The default value is 100.

Request Lenses

describeBatchPredictions_eq :: Lens' DescribeBatchPredictions (Maybe Text)

The equal to operator. The BatchPrediction results will have FilterVariable values that exactly match the value specified with EQ.

describeBatchPredictions_nextToken :: Lens' DescribeBatchPredictions (Maybe Text)

An ID of the page in the paginated results.

describeBatchPredictions_sortOrder :: Lens' DescribeBatchPredictions (Maybe SortOrder)

A two-value parameter that determines the sequence of the resulting list of MLModels.

asc - Arranges the list in ascending order (A-Z, 0-9).

dsc - Arranges the list in descending order (Z-A, 9-0).

Results are sorted by FilterVariable.

describeBatchPredictions_filterVariable :: Lens' DescribeBatchPredictions (Maybe BatchPredictionFilterVariable)

Use one of the following variables to filter a list of BatchPrediction:

CreatedAt - Sets the search criteria to the BatchPrediction creation date.

Status - Sets the search criteria to the BatchPrediction status.

Name - Sets the search criteria to the contents of the BatchPrediction ____ Name.

IAMUser - Sets the search criteria to the user account that invoked the BatchPrediction creation.

MLModelId - Sets the search criteria to the MLModel used in the BatchPrediction.

DataSourceId - Sets the search criteria to the DataSource used in the BatchPrediction.

DataURI - Sets the search criteria to the data file(s) used in the BatchPrediction. The URL can identify either a file or an Amazon Simple Storage Solution (Amazon S3) bucket or directory.

describeBatchPredictions_gt :: Lens' DescribeBatchPredictions (Maybe Text)

The greater than operator. The BatchPrediction results will have FilterVariable values that are greater than the value specified with GT.

describeBatchPredictions_ne :: Lens' DescribeBatchPredictions (Maybe Text)

The not equal to operator. The BatchPrediction results will have FilterVariable values not equal to the value specified with NE.

describeBatchPredictions_prefix :: Lens' DescribeBatchPredictions (Maybe Text)

A string that is found at the beginning of a variable, such as Name or Id.

For example, a Batch Prediction operation could have the Name 2014-09-09-HolidayGiftMailer. To search for this BatchPrediction, select Name for the FilterVariable and any of the following strings for the Prefix:

2014-09

2014-09-09

2014-09-09-Holiday

describeBatchPredictions_ge :: Lens' DescribeBatchPredictions (Maybe Text)

The greater than or equal to operator. The BatchPrediction results will have FilterVariable values that are greater than or equal to the value specified with GE.

describeBatchPredictions_le :: Lens' DescribeBatchPredictions (Maybe Text)

The less than or equal to operator. The BatchPrediction results will have FilterVariable values that are less than or equal to the value specified with LE.

describeBatchPredictions_lt :: Lens' DescribeBatchPredictions (Maybe Text)

The less than operator. The BatchPrediction results will have FilterVariable values that are less than the value specified with LT.

describeBatchPredictions_limit :: Lens' DescribeBatchPredictions (Maybe Natural)

The number of pages of information to include in the result. The range of acceptable values is 1 through 100. The default value is 100.

Destructuring the Response

data DescribeBatchPredictionsResponse

Represents the output of a DescribeBatchPredictions operation. The content is essentially a list of BatchPredictions.

See: newDescribeBatchPredictionsResponse smart constructor.

Constructors

DescribeBatchPredictionsResponse'

Fields

nextToken :: Maybe Text

The ID of the next page in the paginated results that indicates at least one more page follows.

results :: Maybe [BatchPrediction]

A list of BatchPrediction objects that meet the search criteria.

httpStatus :: Int

The response's http status code.

newDescribeBatchPredictionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeBatchPredictionsResponse'

-> DescribeBatchPredictionsResponse

Create a value of DescribeBatchPredictionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeBatchPredictions', describeBatchPredictionsResponse_nextToken - The ID of the next page in the paginated results that indicates at least one more page follows.

$sel:results:DescribeBatchPredictionsResponse', describeBatchPredictionsResponse_results - A list of BatchPrediction objects that meet the search criteria.

$sel:httpStatus:DescribeBatchPredictionsResponse', describeBatchPredictionsResponse_httpStatus - The response's http status code.

Response Lenses

describeBatchPredictionsResponse_nextToken :: Lens' DescribeBatchPredictionsResponse (Maybe Text)

The ID of the next page in the paginated results that indicates at least one more page follows.

describeBatchPredictionsResponse_results :: Lens' DescribeBatchPredictionsResponse (Maybe [BatchPrediction])

A list of BatchPrediction objects that meet the search criteria.

describeBatchPredictionsResponse_httpStatus :: Lens' DescribeBatchPredictionsResponse Int

The response's http status code.