Network.AWS.MachineLearning.DescribeEvaluations

Description

Returns a list of DescribeEvaluations that match the search criteria in the request.

This operation returns paginated results.

Creating a Request

data DescribeEvaluations

See: newDescribeEvaluations smart constructor.

Constructors

DescribeEvaluations'

Fields

eq :: Maybe Text

The equal to operator. The Evaluation results will have FilterVariable values that exactly match the value specified with EQ.

ge :: Maybe Text

The greater than or equal to operator. The Evaluation results will have FilterVariable values that are greater than or equal to the value specified with GE.

prefix :: Maybe Text

A string that is found at the beginning of a variable, such as Name or Id.

For example, an Evaluation could have the Name 2014-09-09-HolidayGiftMailer. To search for this Evaluation, select Name for the FilterVariable and any of the following strings for the Prefix:

2014-09

2014-09-09

2014-09-09-Holiday

gt :: Maybe Text

The greater than operator. The Evaluation results will have FilterVariable values that are greater than the value specified with GT.

ne :: Maybe Text

The not equal to operator. The Evaluation results will have FilterVariable values not equal to the value specified with NE.

nextToken :: Maybe Text

The ID of the page in the paginated results.

sortOrder :: Maybe SortOrder

A two-value parameter that determines the sequence of the resulting list of Evaluation.

asc - Arranges the list in ascending order (A-Z, 0-9).

dsc - Arranges the list in descending order (Z-A, 9-0).

Results are sorted by FilterVariable.

limit :: Maybe Natural

The maximum number of Evaluation to include in the result.

lt :: Maybe Text

The less than operator. The Evaluation results will have FilterVariable values that are less than the value specified with LT.

filterVariable :: Maybe EvaluationFilterVariable

Use one of the following variable to filter a list of Evaluation objects:

CreatedAt - Sets the search criteria to the Evaluation creation date.

Status - Sets the search criteria to the Evaluation status.

Name - Sets the search criteria to the contents of Evaluation ____ Name.

IAMUser - Sets the search criteria to the user account that invoked an Evaluation.

MLModelId - Sets the search criteria to the MLModel that was evaluated.

DataSourceId - Sets the search criteria to the DataSource used in Evaluation.

DataUri - Sets the search criteria to the data file(s) used in Evaluation. The URL can identify either a file or an Amazon Simple Storage Solution (Amazon S3) bucket or directory.

le :: Maybe Text

The less than or equal to operator. The Evaluation results will have FilterVariable values that are less than or equal to the value specified with LE.

newDescribeEvaluations :: DescribeEvaluations

Create a value of DescribeEvaluations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eq:DescribeEvaluations', describeEvaluations_eq - The equal to operator. The Evaluation results will have FilterVariable values that exactly match the value specified with EQ.

$sel:ge:DescribeEvaluations', describeEvaluations_ge - The greater than or equal to operator. The Evaluation results will have FilterVariable values that are greater than or equal to the value specified with GE.

$sel:prefix:DescribeEvaluations', describeEvaluations_prefix - A string that is found at the beginning of a variable, such as Name or Id.

For example, an Evaluation could have the Name 2014-09-09-HolidayGiftMailer. To search for this Evaluation, select Name for the FilterVariable and any of the following strings for the Prefix:

2014-09

2014-09-09

2014-09-09-Holiday

$sel:gt:DescribeEvaluations', describeEvaluations_gt - The greater than operator. The Evaluation results will have FilterVariable values that are greater than the value specified with GT.

$sel:ne:DescribeEvaluations', describeEvaluations_ne - The not equal to operator. The Evaluation results will have FilterVariable values not equal to the value specified with NE.

$sel:nextToken:DescribeEvaluations', describeEvaluations_nextToken - The ID of the page in the paginated results.

$sel:sortOrder:DescribeEvaluations', describeEvaluations_sortOrder - A two-value parameter that determines the sequence of the resulting list of Evaluation.

asc - Arranges the list in ascending order (A-Z, 0-9).

dsc - Arranges the list in descending order (Z-A, 9-0).

Results are sorted by FilterVariable.

$sel:limit:DescribeEvaluations', describeEvaluations_limit - The maximum number of Evaluation to include in the result.

$sel:lt:DescribeEvaluations', describeEvaluations_lt - The less than operator. The Evaluation results will have FilterVariable values that are less than the value specified with LT.

$sel:filterVariable:DescribeEvaluations', describeEvaluations_filterVariable - Use one of the following variable to filter a list of Evaluation objects:

CreatedAt - Sets the search criteria to the Evaluation creation date.

Status - Sets the search criteria to the Evaluation status.

Name - Sets the search criteria to the contents of Evaluation ____ Name.

IAMUser - Sets the search criteria to the user account that invoked an Evaluation.

MLModelId - Sets the search criteria to the MLModel that was evaluated.

DataSourceId - Sets the search criteria to the DataSource used in Evaluation.

DataUri - Sets the search criteria to the data file(s) used in Evaluation. The URL can identify either a file or an Amazon Simple Storage Solution (Amazon S3) bucket or directory.

$sel:le:DescribeEvaluations', describeEvaluations_le - The less than or equal to operator. The Evaluation results will have FilterVariable values that are less than or equal to the value specified with LE.

Request Lenses

describeEvaluations_eq :: Lens' DescribeEvaluations (Maybe Text)

The equal to operator. The Evaluation results will have FilterVariable values that exactly match the value specified with EQ.

describeEvaluations_ge :: Lens' DescribeEvaluations (Maybe Text)

The greater than or equal to operator. The Evaluation results will have FilterVariable values that are greater than or equal to the value specified with GE.

describeEvaluations_prefix :: Lens' DescribeEvaluations (Maybe Text)

A string that is found at the beginning of a variable, such as Name or Id.

For example, an Evaluation could have the Name 2014-09-09-HolidayGiftMailer. To search for this Evaluation, select Name for the FilterVariable and any of the following strings for the Prefix:

2014-09

2014-09-09

2014-09-09-Holiday

describeEvaluations_gt :: Lens' DescribeEvaluations (Maybe Text)

The greater than operator. The Evaluation results will have FilterVariable values that are greater than the value specified with GT.

describeEvaluations_ne :: Lens' DescribeEvaluations (Maybe Text)

The not equal to operator. The Evaluation results will have FilterVariable values not equal to the value specified with NE.

describeEvaluations_nextToken :: Lens' DescribeEvaluations (Maybe Text)

The ID of the page in the paginated results.

describeEvaluations_sortOrder :: Lens' DescribeEvaluations (Maybe SortOrder)

A two-value parameter that determines the sequence of the resulting list of Evaluation.

asc - Arranges the list in ascending order (A-Z, 0-9).

dsc - Arranges the list in descending order (Z-A, 9-0).

Results are sorted by FilterVariable.

describeEvaluations_limit :: Lens' DescribeEvaluations (Maybe Natural)

The maximum number of Evaluation to include in the result.

describeEvaluations_lt :: Lens' DescribeEvaluations (Maybe Text)

The less than operator. The Evaluation results will have FilterVariable values that are less than the value specified with LT.

describeEvaluations_filterVariable :: Lens' DescribeEvaluations (Maybe EvaluationFilterVariable)

Use one of the following variable to filter a list of Evaluation objects:

CreatedAt - Sets the search criteria to the Evaluation creation date.

Status - Sets the search criteria to the Evaluation status.

Name - Sets the search criteria to the contents of Evaluation ____ Name.

IAMUser - Sets the search criteria to the user account that invoked an Evaluation.

MLModelId - Sets the search criteria to the MLModel that was evaluated.

DataSourceId - Sets the search criteria to the DataSource used in Evaluation.

DataUri - Sets the search criteria to the data file(s) used in Evaluation. The URL can identify either a file or an Amazon Simple Storage Solution (Amazon S3) bucket or directory.

describeEvaluations_le :: Lens' DescribeEvaluations (Maybe Text)

The less than or equal to operator. The Evaluation results will have FilterVariable values that are less than or equal to the value specified with LE.

Destructuring the Response

data DescribeEvaluationsResponse

Represents the query results from a DescribeEvaluations operation. The content is essentially a list of Evaluation.

See: newDescribeEvaluationsResponse smart constructor.

Constructors

DescribeEvaluationsResponse'

Fields

results :: Maybe [Evaluation]

A list of Evaluation that meet the search criteria.

nextToken :: Maybe Text

The ID of the next page in the paginated results that indicates at least one more page follows.

httpStatus :: Int

The response's http status code.

newDescribeEvaluationsResponse

Arguments

:: Int
$sel:httpStatus:DescribeEvaluationsResponse'

-> DescribeEvaluationsResponse

Create a value of DescribeEvaluationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:results:DescribeEvaluationsResponse', describeEvaluationsResponse_results - A list of Evaluation that meet the search criteria.

$sel:nextToken:DescribeEvaluations', describeEvaluationsResponse_nextToken - The ID of the next page in the paginated results that indicates at least one more page follows.

$sel:httpStatus:DescribeEvaluationsResponse', describeEvaluationsResponse_httpStatus - The response's http status code.

Response Lenses

describeEvaluationsResponse_results :: Lens' DescribeEvaluationsResponse (Maybe [Evaluation])

A list of Evaluation that meet the search criteria.

describeEvaluationsResponse_nextToken :: Lens' DescribeEvaluationsResponse (Maybe Text)

The ID of the next page in the paginated results that indicates at least one more page follows.

describeEvaluationsResponse_httpStatus :: Lens' DescribeEvaluationsResponse Int

The response's http status code.