Network.AWS.MachineLearning.DescribeMLModels

Description

Returns a list of MLModel that match the search criteria in the request.

This operation returns paginated results.

Creating a Request

data DescribeMLModels

See: newDescribeMLModels smart constructor.

Constructors

DescribeMLModels'

Fields

eq :: Maybe Text

The equal to operator. The MLModel results will have FilterVariable values that exactly match the value specified with EQ.

nextToken :: Maybe Text

The ID of the page in the paginated results.

sortOrder :: Maybe SortOrder

A two-value parameter that determines the sequence of the resulting list of MLModel.

asc - Arranges the list in ascending order (A-Z, 0-9).

dsc - Arranges the list in descending order (Z-A, 9-0).

Results are sorted by FilterVariable.

filterVariable :: Maybe MLModelFilterVariable

Use one of the following variables to filter a list of MLModel:

CreatedAt - Sets the search criteria to MLModel creation date.

Status - Sets the search criteria to MLModel status.

Name - Sets the search criteria to the contents of MLModel ____ Name.

IAMUser - Sets the search criteria to the user account that invoked the MLModel creation.

TrainingDataSourceId - Sets the search criteria to the DataSource used to train one or more MLModel.

RealtimeEndpointStatus - Sets the search criteria to the MLModel real-time endpoint status.

MLModelType - Sets the search criteria to MLModel type: binary, regression, or multi-class.

Algorithm - Sets the search criteria to the algorithm that the MLModel uses.

TrainingDataURI - Sets the search criteria to the data file(s) used in training a MLModel. The URL can identify either a file or an Amazon Simple Storage Service (Amazon S3) bucket or directory.

gt :: Maybe Text

The greater than operator. The MLModel results will have FilterVariable values that are greater than the value specified with GT.

ne :: Maybe Text

The not equal to operator. The MLModel results will have FilterVariable values not equal to the value specified with NE.

prefix :: Maybe Text

A string that is found at the beginning of a variable, such as Name or Id.

For example, an MLModel could have the Name 2014-09-09-HolidayGiftMailer. To search for this MLModel, select Name for the FilterVariable and any of the following strings for the Prefix:

2014-09

2014-09-09

2014-09-09-Holiday

ge :: Maybe Text

The greater than or equal to operator. The MLModel results will have FilterVariable values that are greater than or equal to the value specified with GE.

le :: Maybe Text

The less than or equal to operator. The MLModel results will have FilterVariable values that are less than or equal to the value specified with LE.

lt :: Maybe Text

The less than operator. The MLModel results will have FilterVariable values that are less than the value specified with LT.

limit :: Maybe Natural

The number of pages of information to include in the result. The range of acceptable values is 1 through 100. The default value is 100.

newDescribeMLModels :: DescribeMLModels

Create a value of DescribeMLModels with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eq:DescribeMLModels', describeMLModels_eq - The equal to operator. The MLModel results will have FilterVariable values that exactly match the value specified with EQ.

$sel:nextToken:DescribeMLModels', describeMLModels_nextToken - The ID of the page in the paginated results.

$sel:sortOrder:DescribeMLModels', describeMLModels_sortOrder - A two-value parameter that determines the sequence of the resulting list of MLModel.

asc - Arranges the list in ascending order (A-Z, 0-9).

dsc - Arranges the list in descending order (Z-A, 9-0).

Results are sorted by FilterVariable.

$sel:filterVariable:DescribeMLModels', describeMLModels_filterVariable - Use one of the following variables to filter a list of MLModel:

CreatedAt - Sets the search criteria to MLModel creation date.

Status - Sets the search criteria to MLModel status.

Name - Sets the search criteria to the contents of MLModel ____ Name.

IAMUser - Sets the search criteria to the user account that invoked the MLModel creation.

TrainingDataSourceId - Sets the search criteria to the DataSource used to train one or more MLModel.

RealtimeEndpointStatus - Sets the search criteria to the MLModel real-time endpoint status.

MLModelType - Sets the search criteria to MLModel type: binary, regression, or multi-class.

Algorithm - Sets the search criteria to the algorithm that the MLModel uses.

TrainingDataURI - Sets the search criteria to the data file(s) used in training a MLModel. The URL can identify either a file or an Amazon Simple Storage Service (Amazon S3) bucket or directory.

$sel:gt:DescribeMLModels', describeMLModels_gt - The greater than operator. The MLModel results will have FilterVariable values that are greater than the value specified with GT.

$sel:ne:DescribeMLModels', describeMLModels_ne - The not equal to operator. The MLModel results will have FilterVariable values not equal to the value specified with NE.

$sel:prefix:DescribeMLModels', describeMLModels_prefix - A string that is found at the beginning of a variable, such as Name or Id.

For example, an MLModel could have the Name 2014-09-09-HolidayGiftMailer. To search for this MLModel, select Name for the FilterVariable and any of the following strings for the Prefix:

2014-09

2014-09-09

2014-09-09-Holiday

$sel:ge:DescribeMLModels', describeMLModels_ge - The greater than or equal to operator. The MLModel results will have FilterVariable values that are greater than or equal to the value specified with GE.

$sel:le:DescribeMLModels', describeMLModels_le - The less than or equal to operator. The MLModel results will have FilterVariable values that are less than or equal to the value specified with LE.

$sel:lt:DescribeMLModels', describeMLModels_lt - The less than operator. The MLModel results will have FilterVariable values that are less than the value specified with LT.

$sel:limit:DescribeMLModels', describeMLModels_limit - The number of pages of information to include in the result. The range of acceptable values is 1 through 100. The default value is 100.

Request Lenses

describeMLModels_eq :: Lens' DescribeMLModels (Maybe Text)

The equal to operator. The MLModel results will have FilterVariable values that exactly match the value specified with EQ.

describeMLModels_nextToken :: Lens' DescribeMLModels (Maybe Text)

The ID of the page in the paginated results.

describeMLModels_sortOrder :: Lens' DescribeMLModels (Maybe SortOrder)

A two-value parameter that determines the sequence of the resulting list of MLModel.

asc - Arranges the list in ascending order (A-Z, 0-9).

dsc - Arranges the list in descending order (Z-A, 9-0).

Results are sorted by FilterVariable.

describeMLModels_filterVariable :: Lens' DescribeMLModels (Maybe MLModelFilterVariable)

Use one of the following variables to filter a list of MLModel:

CreatedAt - Sets the search criteria to MLModel creation date.

Status - Sets the search criteria to MLModel status.

Name - Sets the search criteria to the contents of MLModel ____ Name.

IAMUser - Sets the search criteria to the user account that invoked the MLModel creation.

TrainingDataSourceId - Sets the search criteria to the DataSource used to train one or more MLModel.

RealtimeEndpointStatus - Sets the search criteria to the MLModel real-time endpoint status.

MLModelType - Sets the search criteria to MLModel type: binary, regression, or multi-class.

Algorithm - Sets the search criteria to the algorithm that the MLModel uses.

TrainingDataURI - Sets the search criteria to the data file(s) used in training a MLModel. The URL can identify either a file or an Amazon Simple Storage Service (Amazon S3) bucket or directory.

describeMLModels_gt :: Lens' DescribeMLModels (Maybe Text)

The greater than operator. The MLModel results will have FilterVariable values that are greater than the value specified with GT.

describeMLModels_ne :: Lens' DescribeMLModels (Maybe Text)

The not equal to operator. The MLModel results will have FilterVariable values not equal to the value specified with NE.

describeMLModels_prefix :: Lens' DescribeMLModels (Maybe Text)

A string that is found at the beginning of a variable, such as Name or Id.

For example, an MLModel could have the Name 2014-09-09-HolidayGiftMailer. To search for this MLModel, select Name for the FilterVariable and any of the following strings for the Prefix:

2014-09

2014-09-09

2014-09-09-Holiday

describeMLModels_ge :: Lens' DescribeMLModels (Maybe Text)

The greater than or equal to operator. The MLModel results will have FilterVariable values that are greater than or equal to the value specified with GE.

describeMLModels_le :: Lens' DescribeMLModels (Maybe Text)

The less than or equal to operator. The MLModel results will have FilterVariable values that are less than or equal to the value specified with LE.

describeMLModels_lt :: Lens' DescribeMLModels (Maybe Text)

The less than operator. The MLModel results will have FilterVariable values that are less than the value specified with LT.

describeMLModels_limit :: Lens' DescribeMLModels (Maybe Natural)

The number of pages of information to include in the result. The range of acceptable values is 1 through 100. The default value is 100.

Destructuring the Response

data DescribeMLModelsResponse

Represents the output of a DescribeMLModels operation. The content is essentially a list of MLModel.

See: newDescribeMLModelsResponse smart constructor.

Constructors

DescribeMLModelsResponse'

Fields

nextToken :: Maybe Text

The ID of the next page in the paginated results that indicates at least one more page follows.

results :: Maybe [MLModel]

A list of MLModel that meet the search criteria.

httpStatus :: Int

The response's http status code.

newDescribeMLModelsResponse

Arguments

:: Int
$sel:httpStatus:DescribeMLModelsResponse'

-> DescribeMLModelsResponse

Create a value of DescribeMLModelsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeMLModels', describeMLModelsResponse_nextToken - The ID of the next page in the paginated results that indicates at least one more page follows.

$sel:results:DescribeMLModelsResponse', describeMLModelsResponse_results - A list of MLModel that meet the search criteria.

$sel:httpStatus:DescribeMLModelsResponse', describeMLModelsResponse_httpStatus - The response's http status code.

Response Lenses

describeMLModelsResponse_nextToken :: Lens' DescribeMLModelsResponse (Maybe Text)

The ID of the next page in the paginated results that indicates at least one more page follows.

describeMLModelsResponse_results :: Lens' DescribeMLModelsResponse (Maybe [MLModel])

A list of MLModel that meet the search criteria.

describeMLModelsResponse_httpStatus :: Lens' DescribeMLModelsResponse Int

The response's http status code.