Network.AWS.MachineLearning.GetDataSource

Description

Returns a DataSource that includes metadata and data file information, as well as the current status of the DataSource.

GetDataSource provides results in normal or verbose format. The verbose format adds the schema description and the list of files pointed to by the DataSource to the normal format.

Creating a Request

data GetDataSource

See: newGetDataSource smart constructor.

Constructors

GetDataSource'

Fields

verbose :: Maybe Bool

Specifies whether the GetDataSource operation should return DataSourceSchema.

If true, DataSourceSchema is returned.

If false, DataSourceSchema is not returned.

dataSourceId :: Text

The ID assigned to the DataSource at creation.

newGetDataSource

Arguments

:: Text
$sel:dataSourceId:GetDataSource'

-> GetDataSource

Create a value of GetDataSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:verbose:GetDataSource', getDataSource_verbose - Specifies whether the GetDataSource operation should return DataSourceSchema.

If true, DataSourceSchema is returned.

If false, DataSourceSchema is not returned.

$sel:dataSourceId:GetDataSource', getDataSource_dataSourceId - The ID assigned to the DataSource at creation.

Request Lenses

getDataSource_verbose :: Lens' GetDataSource (Maybe Bool)

Specifies whether the GetDataSource operation should return DataSourceSchema.

If true, DataSourceSchema is returned.

If false, DataSourceSchema is not returned.

getDataSource_dataSourceId :: Lens' GetDataSource Text

The ID assigned to the DataSource at creation.

Destructuring the Response

data GetDataSourceResponse

Represents the output of a GetDataSource operation and describes a DataSource.

See: newGetDataSourceResponse smart constructor.

Constructors

GetDataSourceResponse'

Fields

status :: Maybe EntityStatus

The current status of the DataSource. This element can have one of the following values:

PENDING - Amazon ML submitted a request to create a DataSource.

INPROGRESS - The creation process is underway.

FAILED - The request to create a DataSource did not run to completion. It is not usable.

COMPLETED - The creation process completed successfully.

DELETED - The DataSource is marked as deleted. It is not usable.

numberOfFiles :: Maybe Integer

The number of data files referenced by the DataSource.

lastUpdatedAt :: Maybe POSIX

The time of the most recent edit to the DataSource. The time is expressed in epoch time.

createdAt :: Maybe POSIX

The time that the DataSource was created. The time is expressed in epoch time.

computeTime :: Maybe Integer

The approximate CPU time in milliseconds that Amazon Machine Learning spent processing the DataSource, normalized and scaled on computation resources. ComputeTime is only available if the DataSource is in the COMPLETED state and the ComputeStatistics is set to true.

dataSourceId :: Maybe Text

The ID assigned to the DataSource at creation. This value should be identical to the value of the DataSourceId in the request.

rDSMetadata :: Maybe RDSMetadata

dataSizeInBytes :: Maybe Integer

The total size of observations in the data files.

dataSourceSchema :: Maybe Text

The schema used by all of the data files of this DataSource.

Note: This parameter is provided as part of the verbose format.

startedAt :: Maybe POSIX

The epoch time when Amazon Machine Learning marked the DataSource as INPROGRESS. StartedAt isn't available if the DataSource is in the PENDING state.

finishedAt :: Maybe POSIX

The epoch time when Amazon Machine Learning marked the DataSource as COMPLETED or FAILED. FinishedAt is only available when the DataSource is in the COMPLETED or FAILED state.

createdByIamUser :: Maybe Text

The AWS user account from which the DataSource was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

name :: Maybe Text

A user-supplied name or description of the DataSource.

logUri :: Maybe Text

A link to the file containing logs of CreateDataSourceFrom* operations.

dataLocationS3 :: Maybe Text

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

computeStatistics :: Maybe Bool

The parameter is true if statistics need to be generated from the observation data.

message :: Maybe Text

The user-supplied description of the most recent details about creating the DataSource.

redshiftMetadata :: Maybe RedshiftMetadata

dataRearrangement :: Maybe Text

A JSON string that represents the splitting and rearrangement requirement used when this DataSource was created.

roleARN :: Maybe Text

httpStatus :: Int

The response's http status code.

newGetDataSourceResponse

Arguments

:: Int
$sel:httpStatus:GetDataSourceResponse'

-> GetDataSourceResponse

Create a value of GetDataSourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetDataSourceResponse', getDataSourceResponse_status - The current status of the DataSource. This element can have one of the following values:

PENDING - Amazon ML submitted a request to create a DataSource.

INPROGRESS - The creation process is underway.

FAILED - The request to create a DataSource did not run to completion. It is not usable.

COMPLETED - The creation process completed successfully.

DELETED - The DataSource is marked as deleted. It is not usable.

$sel:numberOfFiles:GetDataSourceResponse', getDataSourceResponse_numberOfFiles - The number of data files referenced by the DataSource.

$sel:lastUpdatedAt:GetDataSourceResponse', getDataSourceResponse_lastUpdatedAt - The time of the most recent edit to the DataSource. The time is expressed in epoch time.

$sel:createdAt:GetDataSourceResponse', getDataSourceResponse_createdAt - The time that the DataSource was created. The time is expressed in epoch time.

$sel:computeTime:GetDataSourceResponse', getDataSourceResponse_computeTime - The approximate CPU time in milliseconds that Amazon Machine Learning spent processing the DataSource, normalized and scaled on computation resources. ComputeTime is only available if the DataSource is in the COMPLETED state and the ComputeStatistics is set to true.

$sel:dataSourceId:GetDataSource', getDataSourceResponse_dataSourceId - The ID assigned to the DataSource at creation. This value should be identical to the value of the DataSourceId in the request.

$sel:rDSMetadata:GetDataSourceResponse', getDataSourceResponse_rDSMetadata - Undocumented member.

$sel:dataSizeInBytes:GetDataSourceResponse', getDataSourceResponse_dataSizeInBytes - The total size of observations in the data files.

$sel:dataSourceSchema:GetDataSourceResponse', getDataSourceResponse_dataSourceSchema - The schema used by all of the data files of this DataSource.

Note: This parameter is provided as part of the verbose format.

$sel:startedAt:GetDataSourceResponse', getDataSourceResponse_startedAt - The epoch time when Amazon Machine Learning marked the DataSource as INPROGRESS. StartedAt isn't available if the DataSource is in the PENDING state.

$sel:finishedAt:GetDataSourceResponse', getDataSourceResponse_finishedAt - The epoch time when Amazon Machine Learning marked the DataSource as COMPLETED or FAILED. FinishedAt is only available when the DataSource is in the COMPLETED or FAILED state.

$sel:createdByIamUser:GetDataSourceResponse', getDataSourceResponse_createdByIamUser - The AWS user account from which the DataSource was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

$sel:name:GetDataSourceResponse', getDataSourceResponse_name - A user-supplied name or description of the DataSource.

$sel:logUri:GetDataSourceResponse', getDataSourceResponse_logUri - A link to the file containing logs of CreateDataSourceFrom* operations.

$sel:dataLocationS3:GetDataSourceResponse', getDataSourceResponse_dataLocationS3 - The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

$sel:computeStatistics:GetDataSourceResponse', getDataSourceResponse_computeStatistics - The parameter is true if statistics need to be generated from the observation data.

$sel:message:GetDataSourceResponse', getDataSourceResponse_message - The user-supplied description of the most recent details about creating the DataSource.

$sel:redshiftMetadata:GetDataSourceResponse', getDataSourceResponse_redshiftMetadata - Undocumented member.

$sel:dataRearrangement:GetDataSourceResponse', getDataSourceResponse_dataRearrangement - A JSON string that represents the splitting and rearrangement requirement used when this DataSource was created.

$sel:roleARN:GetDataSourceResponse', getDataSourceResponse_roleARN - Undocumented member.

$sel:httpStatus:GetDataSourceResponse', getDataSourceResponse_httpStatus - The response's http status code.

Response Lenses

getDataSourceResponse_status :: Lens' GetDataSourceResponse (Maybe EntityStatus)

The current status of the DataSource. This element can have one of the following values:

PENDING - Amazon ML submitted a request to create a DataSource.

INPROGRESS - The creation process is underway.

FAILED - The request to create a DataSource did not run to completion. It is not usable.

COMPLETED - The creation process completed successfully.

DELETED - The DataSource is marked as deleted. It is not usable.

getDataSourceResponse_numberOfFiles :: Lens' GetDataSourceResponse (Maybe Integer)

The number of data files referenced by the DataSource.

getDataSourceResponse_lastUpdatedAt :: Lens' GetDataSourceResponse (Maybe UTCTime)

The time of the most recent edit to the DataSource. The time is expressed in epoch time.

getDataSourceResponse_createdAt :: Lens' GetDataSourceResponse (Maybe UTCTime)

The time that the DataSource was created. The time is expressed in epoch time.

getDataSourceResponse_computeTime :: Lens' GetDataSourceResponse (Maybe Integer)

The approximate CPU time in milliseconds that Amazon Machine Learning spent processing the DataSource, normalized and scaled on computation resources. ComputeTime is only available if the DataSource is in the COMPLETED state and the ComputeStatistics is set to true.

getDataSourceResponse_dataSourceId :: Lens' GetDataSourceResponse (Maybe Text)

The ID assigned to the DataSource at creation. This value should be identical to the value of the DataSourceId in the request.

getDataSourceResponse_rDSMetadata :: Lens' GetDataSourceResponse (Maybe RDSMetadata)

Undocumented member.

getDataSourceResponse_dataSizeInBytes :: Lens' GetDataSourceResponse (Maybe Integer)

The total size of observations in the data files.

getDataSourceResponse_dataSourceSchema :: Lens' GetDataSourceResponse (Maybe Text)

The schema used by all of the data files of this DataSource.

Note: This parameter is provided as part of the verbose format.

getDataSourceResponse_startedAt :: Lens' GetDataSourceResponse (Maybe UTCTime)

The epoch time when Amazon Machine Learning marked the DataSource as INPROGRESS. StartedAt isn't available if the DataSource is in the PENDING state.

getDataSourceResponse_finishedAt :: Lens' GetDataSourceResponse (Maybe UTCTime)

The epoch time when Amazon Machine Learning marked the DataSource as COMPLETED or FAILED. FinishedAt is only available when the DataSource is in the COMPLETED or FAILED state.

getDataSourceResponse_createdByIamUser :: Lens' GetDataSourceResponse (Maybe Text)

The AWS user account from which the DataSource was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

getDataSourceResponse_name :: Lens' GetDataSourceResponse (Maybe Text)

A user-supplied name or description of the DataSource.

getDataSourceResponse_logUri :: Lens' GetDataSourceResponse (Maybe Text)

A link to the file containing logs of CreateDataSourceFrom* operations.

getDataSourceResponse_dataLocationS3 :: Lens' GetDataSourceResponse (Maybe Text)

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

getDataSourceResponse_computeStatistics :: Lens' GetDataSourceResponse (Maybe Bool)

The parameter is true if statistics need to be generated from the observation data.

getDataSourceResponse_message :: Lens' GetDataSourceResponse (Maybe Text)

The user-supplied description of the most recent details about creating the DataSource.

getDataSourceResponse_redshiftMetadata :: Lens' GetDataSourceResponse (Maybe RedshiftMetadata)

Undocumented member.

getDataSourceResponse_dataRearrangement :: Lens' GetDataSourceResponse (Maybe Text)

A JSON string that represents the splitting and rearrangement requirement used when this DataSource was created.

getDataSourceResponse_roleARN :: Lens' GetDataSourceResponse (Maybe Text)

Undocumented member.

getDataSourceResponse_httpStatus :: Lens' GetDataSourceResponse Int

The response's http status code.