Network.AWS.MachineLearning.GetMLModel

Description

Returns an MLModel that includes detailed metadata, data source information, and the current status of the MLModel.

GetMLModel provides results in normal or verbose format.

Creating a Request

data GetMLModel

See: newGetMLModel smart constructor.

Constructors

GetMLModel'

Fields

verbose :: Maybe Bool

Specifies whether the GetMLModel operation should return Recipe.

If true, Recipe is returned.

If false, Recipe is not returned.

mLModelId :: Text

The ID assigned to the MLModel at creation.

newGetMLModel

Arguments

:: Text
$sel:mLModelId:GetMLModel'

-> GetMLModel

Create a value of GetMLModel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:verbose:GetMLModel', getMLModel_verbose - Specifies whether the GetMLModel operation should return Recipe.

If true, Recipe is returned.

If false, Recipe is not returned.

$sel:mLModelId:GetMLModel', getMLModel_mLModelId - The ID assigned to the MLModel at creation.

Request Lenses

getMLModel_verbose :: Lens' GetMLModel (Maybe Bool)

Specifies whether the GetMLModel operation should return Recipe.

If true, Recipe is returned.

If false, Recipe is not returned.

getMLModel_mLModelId :: Lens' GetMLModel Text

The ID assigned to the MLModel at creation.

Destructuring the Response

data GetMLModelResponse

Represents the output of a GetMLModel operation, and provides detailed information about a MLModel.

See: newGetMLModelResponse smart constructor.

Constructors

GetMLModelResponse'

Fields

status :: Maybe EntityStatus

The current status of the MLModel. This element can have one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to describe a MLModel.

INPROGRESS - The request is processing.

FAILED - The request did not run to completion. The ML model isn't usable.

COMPLETED - The request completed successfully.

DELETED - The MLModel is marked as deleted. It isn't usable.

startedAt :: Maybe POSIX

The epoch time when Amazon Machine Learning marked the MLModel as INPROGRESS. StartedAt isn't available if the MLModel is in the PENDING state.

schema :: Maybe Text

The schema used by all of the data files referenced by the DataSource.

Note: This parameter is provided as part of the verbose format.

message :: Maybe Text

A description of the most recent details about accessing the MLModel.

recipe :: Maybe Text

The recipe to use when training the MLModel. The Recipe provides detailed information about the observation data to use during training, and manipulations to perform on the observation data during training.

Note: This parameter is provided as part of the verbose format.

createdAt :: Maybe POSIX

The time that the MLModel was created. The time is expressed in epoch time.

scoreThresholdLastUpdatedAt :: Maybe POSIX

The time of the most recent edit to the ScoreThreshold. The time is expressed in epoch time.

endpointInfo :: Maybe RealtimeEndpointInfo

The current endpoint of the MLModel

trainingParameters :: Maybe (HashMap Text Text)

A list of the training parameters in the MLModel. The list is implemented as a map of key-value pairs.

The following is the current set of training parameters:

sgd.maxMLModelSizeInBytes - The maximum allowed size of the model. Depending on the input data, the size of the model might affect its performance.

The value is an integer that ranges from 100000 to 2147483648. The default value is 33554432.

sgd.maxPasses - The number of times that the training process traverses the observations to build the MLModel. The value is an integer that ranges from 1 to 10000. The default value is 10.

sgd.shuffleType - Whether Amazon ML shuffles the training data. Shuffling data improves a model's ability to find the optimal solution for a variety of data types. The valid values are auto and none. The default value is none. We strongly recommend that you shuffle your data.

sgd.l1RegularizationAmount - The coefficient regularization L1 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to zero, resulting in a sparse feature set. If you use this parameter, start by specifying a small value, such as 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is to not use L1 normalization. This parameter can't be used when L2 is specified. Use this parameter sparingly.

sgd.l2RegularizationAmount - The coefficient regularization L2 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to small, nonzero values. If you use this parameter, start by specifying a small value, such as 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is to not use L2 normalization. This parameter can't be used when L1 is specified. Use this parameter sparingly.

name :: Maybe Text

A user-supplied name or description of the MLModel.

createdByIamUser :: Maybe Text

The AWS user account from which the MLModel was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

finishedAt :: Maybe POSIX

The epoch time when Amazon Machine Learning marked the MLModel as COMPLETED or FAILED. FinishedAt is only available when the MLModel is in the COMPLETED or FAILED state.

scoreThreshold :: Maybe Double

The scoring threshold is used in binary classification MLModel models. It marks the boundary between a positive prediction and a negative prediction.

Output values greater than or equal to the threshold receive a positive result from the MLModel, such as true. Output values less than the threshold receive a negative response from the MLModel, such as false.

mLModelType :: Maybe MLModelType

Identifies the MLModel category. The following are the available types:

REGRESSION -- Produces a numeric result. For example, "What price should a house be listed at?"

BINARY -- Produces one of two possible results. For example, "Is this an e-commerce website?"

MULTICLASS -- Produces one of several possible results. For example, "Is this a HIGH, LOW or MEDIUM risk trade?"

sizeInBytes :: Maybe Integer

mLModelId :: Maybe Text

The MLModel ID, which is same as the MLModelId in the request.

inputDataLocationS3 :: Maybe Text

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

computeTime :: Maybe Integer

The approximate CPU time in milliseconds that Amazon Machine Learning spent processing the MLModel, normalized and scaled on computation resources. ComputeTime is only available if the MLModel is in the COMPLETED state.

trainingDataSourceId :: Maybe Text

The ID of the training DataSource.

lastUpdatedAt :: Maybe POSIX

The time of the most recent edit to the MLModel. The time is expressed in epoch time.

logUri :: Maybe Text

A link to the file that contains logs of the CreateMLModel operation.

httpStatus :: Int

The response's http status code.

newGetMLModelResponse

Arguments

:: Int
$sel:httpStatus:GetMLModelResponse'

-> GetMLModelResponse

Create a value of GetMLModelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetMLModelResponse', getMLModelResponse_status - The current status of the MLModel. This element can have one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to describe a MLModel.

INPROGRESS - The request is processing.

FAILED - The request did not run to completion. The ML model isn't usable.

COMPLETED - The request completed successfully.

DELETED - The MLModel is marked as deleted. It isn't usable.

$sel:startedAt:GetMLModelResponse', getMLModelResponse_startedAt - The epoch time when Amazon Machine Learning marked the MLModel as INPROGRESS. StartedAt isn't available if the MLModel is in the PENDING state.

$sel:schema:GetMLModelResponse', getMLModelResponse_schema - The schema used by all of the data files referenced by the DataSource.

Note: This parameter is provided as part of the verbose format.

$sel:message:GetMLModelResponse', getMLModelResponse_message - A description of the most recent details about accessing the MLModel.

$sel:recipe:GetMLModelResponse', getMLModelResponse_recipe - The recipe to use when training the MLModel. The Recipe provides detailed information about the observation data to use during training, and manipulations to perform on the observation data during training.

Note: This parameter is provided as part of the verbose format.

$sel:createdAt:GetMLModelResponse', getMLModelResponse_createdAt - The time that the MLModel was created. The time is expressed in epoch time.

$sel:scoreThresholdLastUpdatedAt:GetMLModelResponse', getMLModelResponse_scoreThresholdLastUpdatedAt - The time of the most recent edit to the ScoreThreshold. The time is expressed in epoch time.

$sel:endpointInfo:GetMLModelResponse', getMLModelResponse_endpointInfo - The current endpoint of the MLModel

$sel:trainingParameters:GetMLModelResponse', getMLModelResponse_trainingParameters - A list of the training parameters in the MLModel. The list is implemented as a map of key-value pairs.

The following is the current set of training parameters:

sgd.maxMLModelSizeInBytes - The maximum allowed size of the model. Depending on the input data, the size of the model might affect its performance.

The value is an integer that ranges from 100000 to 2147483648. The default value is 33554432.

sgd.maxPasses - The number of times that the training process traverses the observations to build the MLModel. The value is an integer that ranges from 1 to 10000. The default value is 10.

sgd.shuffleType - Whether Amazon ML shuffles the training data. Shuffling data improves a model's ability to find the optimal solution for a variety of data types. The valid values are auto and none. The default value is none. We strongly recommend that you shuffle your data.

sgd.l1RegularizationAmount - The coefficient regularization L1 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to zero, resulting in a sparse feature set. If you use this parameter, start by specifying a small value, such as 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is to not use L1 normalization. This parameter can't be used when L2 is specified. Use this parameter sparingly.

sgd.l2RegularizationAmount - The coefficient regularization L2 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to small, nonzero values. If you use this parameter, start by specifying a small value, such as 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is to not use L2 normalization. This parameter can't be used when L1 is specified. Use this parameter sparingly.

$sel:name:GetMLModelResponse', getMLModelResponse_name - A user-supplied name or description of the MLModel.

$sel:createdByIamUser:GetMLModelResponse', getMLModelResponse_createdByIamUser - The AWS user account from which the MLModel was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

$sel:finishedAt:GetMLModelResponse', getMLModelResponse_finishedAt - The epoch time when Amazon Machine Learning marked the MLModel as COMPLETED or FAILED. FinishedAt is only available when the MLModel is in the COMPLETED or FAILED state.

$sel:scoreThreshold:GetMLModelResponse', getMLModelResponse_scoreThreshold - The scoring threshold is used in binary classification MLModel models. It marks the boundary between a positive prediction and a negative prediction.

Output values greater than or equal to the threshold receive a positive result from the MLModel, such as true. Output values less than the threshold receive a negative response from the MLModel, such as false.

$sel:mLModelType:GetMLModelResponse', getMLModelResponse_mLModelType - Identifies the MLModel category. The following are the available types:

REGRESSION -- Produces a numeric result. For example, "What price should a house be listed at?"

BINARY -- Produces one of two possible results. For example, "Is this an e-commerce website?"

MULTICLASS -- Produces one of several possible results. For example, "Is this a HIGH, LOW or MEDIUM risk trade?"

$sel:sizeInBytes:GetMLModelResponse', getMLModelResponse_sizeInBytes - Undocumented member.

$sel:mLModelId:GetMLModel', getMLModelResponse_mLModelId - The MLModel ID, which is same as the MLModelId in the request.

$sel:inputDataLocationS3:GetMLModelResponse', getMLModelResponse_inputDataLocationS3 - The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

$sel:computeTime:GetMLModelResponse', getMLModelResponse_computeTime - The approximate CPU time in milliseconds that Amazon Machine Learning spent processing the MLModel, normalized and scaled on computation resources. ComputeTime is only available if the MLModel is in the COMPLETED state.

$sel:trainingDataSourceId:GetMLModelResponse', getMLModelResponse_trainingDataSourceId - The ID of the training DataSource.

$sel:lastUpdatedAt:GetMLModelResponse', getMLModelResponse_lastUpdatedAt - The time of the most recent edit to the MLModel. The time is expressed in epoch time.

$sel:logUri:GetMLModelResponse', getMLModelResponse_logUri - A link to the file that contains logs of the CreateMLModel operation.

$sel:httpStatus:GetMLModelResponse', getMLModelResponse_httpStatus - The response's http status code.

Response Lenses

getMLModelResponse_status :: Lens' GetMLModelResponse (Maybe EntityStatus)

The current status of the MLModel. This element can have one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to describe a MLModel.

INPROGRESS - The request is processing.

FAILED - The request did not run to completion. The ML model isn't usable.

COMPLETED - The request completed successfully.

DELETED - The MLModel is marked as deleted. It isn't usable.

getMLModelResponse_startedAt :: Lens' GetMLModelResponse (Maybe UTCTime)

The epoch time when Amazon Machine Learning marked the MLModel as INPROGRESS. StartedAt isn't available if the MLModel is in the PENDING state.

getMLModelResponse_schema :: Lens' GetMLModelResponse (Maybe Text)

The schema used by all of the data files referenced by the DataSource.

Note: This parameter is provided as part of the verbose format.

getMLModelResponse_message :: Lens' GetMLModelResponse (Maybe Text)

A description of the most recent details about accessing the MLModel.

getMLModelResponse_recipe :: Lens' GetMLModelResponse (Maybe Text)

The recipe to use when training the MLModel. The Recipe provides detailed information about the observation data to use during training, and manipulations to perform on the observation data during training.

Note: This parameter is provided as part of the verbose format.

getMLModelResponse_createdAt :: Lens' GetMLModelResponse (Maybe UTCTime)

The time that the MLModel was created. The time is expressed in epoch time.

getMLModelResponse_scoreThresholdLastUpdatedAt :: Lens' GetMLModelResponse (Maybe UTCTime)

The time of the most recent edit to the ScoreThreshold. The time is expressed in epoch time.

getMLModelResponse_endpointInfo :: Lens' GetMLModelResponse (Maybe RealtimeEndpointInfo)

The current endpoint of the MLModel

getMLModelResponse_trainingParameters :: Lens' GetMLModelResponse (Maybe (HashMap Text Text))

A list of the training parameters in the MLModel. The list is implemented as a map of key-value pairs.

The following is the current set of training parameters:

sgd.maxMLModelSizeInBytes - The maximum allowed size of the model. Depending on the input data, the size of the model might affect its performance.

The value is an integer that ranges from 100000 to 2147483648. The default value is 33554432.

sgd.maxPasses - The number of times that the training process traverses the observations to build the MLModel. The value is an integer that ranges from 1 to 10000. The default value is 10.

sgd.shuffleType - Whether Amazon ML shuffles the training data. Shuffling data improves a model's ability to find the optimal solution for a variety of data types. The valid values are auto and none. The default value is none. We strongly recommend that you shuffle your data.

sgd.l1RegularizationAmount - The coefficient regularization L1 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to zero, resulting in a sparse feature set. If you use this parameter, start by specifying a small value, such as 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is to not use L1 normalization. This parameter can't be used when L2 is specified. Use this parameter sparingly.

sgd.l2RegularizationAmount - The coefficient regularization L2 norm. It controls overfitting the data by penalizing large coefficients. This tends to drive coefficients to small, nonzero values. If you use this parameter, start by specifying a small value, such as 1.0E-08.

The value is a double that ranges from 0 to MAX_DOUBLE. The default is to not use L2 normalization. This parameter can't be used when L1 is specified. Use this parameter sparingly.

getMLModelResponse_name :: Lens' GetMLModelResponse (Maybe Text)

A user-supplied name or description of the MLModel.

getMLModelResponse_createdByIamUser :: Lens' GetMLModelResponse (Maybe Text)

The AWS user account from which the MLModel was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

getMLModelResponse_finishedAt :: Lens' GetMLModelResponse (Maybe UTCTime)

The epoch time when Amazon Machine Learning marked the MLModel as COMPLETED or FAILED. FinishedAt is only available when the MLModel is in the COMPLETED or FAILED state.

getMLModelResponse_scoreThreshold :: Lens' GetMLModelResponse (Maybe Double)

The scoring threshold is used in binary classification MLModel models. It marks the boundary between a positive prediction and a negative prediction.

Output values greater than or equal to the threshold receive a positive result from the MLModel, such as true. Output values less than the threshold receive a negative response from the MLModel, such as false.

getMLModelResponse_mLModelType :: Lens' GetMLModelResponse (Maybe MLModelType)

Identifies the MLModel category. The following are the available types:

REGRESSION -- Produces a numeric result. For example, "What price should a house be listed at?"

BINARY -- Produces one of two possible results. For example, "Is this an e-commerce website?"

MULTICLASS -- Produces one of several possible results. For example, "Is this a HIGH, LOW or MEDIUM risk trade?"

getMLModelResponse_sizeInBytes :: Lens' GetMLModelResponse (Maybe Integer)

Undocumented member.

getMLModelResponse_mLModelId :: Lens' GetMLModelResponse (Maybe Text)

The MLModel ID, which is same as the MLModelId in the request.

getMLModelResponse_inputDataLocationS3 :: Lens' GetMLModelResponse (Maybe Text)

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

getMLModelResponse_computeTime :: Lens' GetMLModelResponse (Maybe Integer)

The approximate CPU time in milliseconds that Amazon Machine Learning spent processing the MLModel, normalized and scaled on computation resources. ComputeTime is only available if the MLModel is in the COMPLETED state.

getMLModelResponse_trainingDataSourceId :: Lens' GetMLModelResponse (Maybe Text)

The ID of the training DataSource.

getMLModelResponse_lastUpdatedAt :: Lens' GetMLModelResponse (Maybe UTCTime)

The time of the most recent edit to the MLModel. The time is expressed in epoch time.

getMLModelResponse_logUri :: Lens' GetMLModelResponse (Maybe Text)

A link to the file that contains logs of the CreateMLModel operation.

getMLModelResponse_httpStatus :: Lens' GetMLModelResponse Int

The response's http status code.