Network.AWS.MachineLearning.Types.BatchPrediction

Description

Documentation

data BatchPrediction

Represents the output of a GetBatchPrediction operation.

The content consists of the detailed metadata, the status, and the data file information of a Batch Prediction.

See: newBatchPrediction smart constructor.

Constructors

BatchPrediction'

Fields

status :: Maybe EntityStatus

The status of the BatchPrediction. This element can have one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to generate predictions for a batch of observations.

INPROGRESS - The process is underway.

FAILED - The request to perform a batch prediction did not run to completion. It is not usable.

COMPLETED - The batch prediction process completed successfully.

DELETED - The BatchPrediction is marked as deleted. It is not usable.

batchPredictionId :: Maybe Text

The ID assigned to the BatchPrediction at creation. This value should be identical to the value of the BatchPredictionID in the request.

startedAt :: Maybe POSIX

outputUri :: Maybe Text

The location of an Amazon S3 bucket or directory to receive the operation results. The following substrings are not allowed in the s3 key portion of the outputURI field: ':', '//', '/./', '/../'.

message :: Maybe Text

A description of the most recent details about processing the batch prediction request.

createdAt :: Maybe POSIX

The time that the BatchPrediction was created. The time is expressed in epoch time.

name :: Maybe Text

A user-supplied name or description of the BatchPrediction.

createdByIamUser :: Maybe Text

The AWS user account that invoked the BatchPrediction. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

finishedAt :: Maybe POSIX

invalidRecordCount :: Maybe Integer

totalRecordCount :: Maybe Integer

batchPredictionDataSourceId :: Maybe Text

The ID of the DataSource that points to the group of observations to predict.

mLModelId :: Maybe Text

The ID of the MLModel that generated predictions for the BatchPrediction request.

inputDataLocationS3 :: Maybe Text

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

computeTime :: Maybe Integer

lastUpdatedAt :: Maybe POSIX

The time of the most recent edit to the BatchPrediction. The time is expressed in epoch time.

newBatchPrediction :: BatchPrediction

Create a value of BatchPrediction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:BatchPrediction', batchPrediction_status - The status of the BatchPrediction. This element can have one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to generate predictions for a batch of observations.

INPROGRESS - The process is underway.

FAILED - The request to perform a batch prediction did not run to completion. It is not usable.

COMPLETED - The batch prediction process completed successfully.

DELETED - The BatchPrediction is marked as deleted. It is not usable.

$sel:batchPredictionId:BatchPrediction', batchPrediction_batchPredictionId - The ID assigned to the BatchPrediction at creation. This value should be identical to the value of the BatchPredictionID in the request.

$sel:startedAt:BatchPrediction', batchPrediction_startedAt - Undocumented member.

$sel:outputUri:BatchPrediction', batchPrediction_outputUri - The location of an Amazon S3 bucket or directory to receive the operation results. The following substrings are not allowed in the s3 key portion of the outputURI field: ':', '//', '/./', '/../'.

$sel:message:BatchPrediction', batchPrediction_message - A description of the most recent details about processing the batch prediction request.

$sel:createdAt:BatchPrediction', batchPrediction_createdAt - The time that the BatchPrediction was created. The time is expressed in epoch time.

$sel:name:BatchPrediction', batchPrediction_name - A user-supplied name or description of the BatchPrediction.

$sel:createdByIamUser:BatchPrediction', batchPrediction_createdByIamUser - The AWS user account that invoked the BatchPrediction. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

$sel:finishedAt:BatchPrediction', batchPrediction_finishedAt - Undocumented member.

$sel:invalidRecordCount:BatchPrediction', batchPrediction_invalidRecordCount - Undocumented member.

$sel:totalRecordCount:BatchPrediction', batchPrediction_totalRecordCount - Undocumented member.

$sel:batchPredictionDataSourceId:BatchPrediction', batchPrediction_batchPredictionDataSourceId - The ID of the DataSource that points to the group of observations to predict.

$sel:mLModelId:BatchPrediction', batchPrediction_mLModelId - The ID of the MLModel that generated predictions for the BatchPrediction request.

$sel:inputDataLocationS3:BatchPrediction', batchPrediction_inputDataLocationS3 - The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

$sel:computeTime:BatchPrediction', batchPrediction_computeTime - Undocumented member.

$sel:lastUpdatedAt:BatchPrediction', batchPrediction_lastUpdatedAt - The time of the most recent edit to the BatchPrediction. The time is expressed in epoch time.

batchPrediction_status :: Lens' BatchPrediction (Maybe EntityStatus)

The status of the BatchPrediction. This element can have one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to generate predictions for a batch of observations.

INPROGRESS - The process is underway.

FAILED - The request to perform a batch prediction did not run to completion. It is not usable.

COMPLETED - The batch prediction process completed successfully.

DELETED - The BatchPrediction is marked as deleted. It is not usable.

batchPrediction_batchPredictionId :: Lens' BatchPrediction (Maybe Text)

The ID assigned to the BatchPrediction at creation. This value should be identical to the value of the BatchPredictionID in the request.

batchPrediction_startedAt :: Lens' BatchPrediction (Maybe UTCTime)

Undocumented member.

batchPrediction_outputUri :: Lens' BatchPrediction (Maybe Text)

The location of an Amazon S3 bucket or directory to receive the operation results. The following substrings are not allowed in the s3 key portion of the outputURI field: ':', '//', '/./', '/../'.

batchPrediction_message :: Lens' BatchPrediction (Maybe Text)

A description of the most recent details about processing the batch prediction request.

batchPrediction_createdAt :: Lens' BatchPrediction (Maybe UTCTime)

The time that the BatchPrediction was created. The time is expressed in epoch time.

batchPrediction_name :: Lens' BatchPrediction (Maybe Text)

A user-supplied name or description of the BatchPrediction.

batchPrediction_createdByIamUser :: Lens' BatchPrediction (Maybe Text)

The AWS user account that invoked the BatchPrediction. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

batchPrediction_finishedAt :: Lens' BatchPrediction (Maybe UTCTime)

Undocumented member.

batchPrediction_invalidRecordCount :: Lens' BatchPrediction (Maybe Integer)

Undocumented member.

batchPrediction_totalRecordCount :: Lens' BatchPrediction (Maybe Integer)

Undocumented member.

batchPrediction_batchPredictionDataSourceId :: Lens' BatchPrediction (Maybe Text)

The ID of the DataSource that points to the group of observations to predict.

batchPrediction_mLModelId :: Lens' BatchPrediction (Maybe Text)

The ID of the MLModel that generated predictions for the BatchPrediction request.

batchPrediction_inputDataLocationS3 :: Lens' BatchPrediction (Maybe Text)

The location of the data file or directory in Amazon Simple Storage Service (Amazon S3).

batchPrediction_computeTime :: Lens' BatchPrediction (Maybe Integer)

Undocumented member.

batchPrediction_lastUpdatedAt :: Lens' BatchPrediction (Maybe UTCTime)

The time of the most recent edit to the BatchPrediction. The time is expressed in epoch time.