Network.AWS.MachineLearning.Types.DataSource

Description

Documentation

data DataSource

Represents the output of the GetDataSource operation.

The content consists of the detailed metadata and data file information and the current status of the DataSource.

See: newDataSource smart constructor.

Constructors

DataSource'

Fields

status :: Maybe EntityStatus

The current status of the DataSource. This element can have one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to create a DataSource.

INPROGRESS - The creation process is underway.

FAILED - The request to create a DataSource did not run to completion. It is not usable.

COMPLETED - The creation process completed successfully.

DELETED - The DataSource is marked as deleted. It is not usable.

startedAt :: Maybe POSIX

roleARN :: Maybe Text

dataRearrangement :: Maybe Text

A JSON string that represents the splitting and rearrangement requirement used when this DataSource was created.

redshiftMetadata :: Maybe RedshiftMetadata

message :: Maybe Text

A description of the most recent details about creating the DataSource.

dataSourceId :: Maybe Text

The ID that is assigned to the DataSource during creation.

computeStatistics :: Maybe Bool

The parameter is true if statistics need to be generated from the observation data.

createdAt :: Maybe POSIX

The time that the DataSource was created. The time is expressed in epoch time.

dataLocationS3 :: Maybe Text

The location and name of the data in Amazon Simple Storage Service (Amazon S3) that is used by a DataSource.

name :: Maybe Text

A user-supplied name or description of the DataSource.

numberOfFiles :: Maybe Integer

The number of data files referenced by the DataSource.

createdByIamUser :: Maybe Text

The AWS user account from which the DataSource was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

finishedAt :: Maybe POSIX

dataSizeInBytes :: Maybe Integer

The total number of observations contained in the data files that the DataSource references.

computeTime :: Maybe Integer

rDSMetadata :: Maybe RDSMetadata

lastUpdatedAt :: Maybe POSIX

The time of the most recent edit to the BatchPrediction. The time is expressed in epoch time.

newDataSource :: DataSource

Create a value of DataSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DataSource', dataSource_status - The current status of the DataSource. This element can have one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to create a DataSource.

INPROGRESS - The creation process is underway.

FAILED - The request to create a DataSource did not run to completion. It is not usable.

COMPLETED - The creation process completed successfully.

DELETED - The DataSource is marked as deleted. It is not usable.

$sel:startedAt:DataSource', dataSource_startedAt - Undocumented member.

$sel:roleARN:DataSource', dataSource_roleARN - Undocumented member.

$sel:dataRearrangement:DataSource', dataSource_dataRearrangement - A JSON string that represents the splitting and rearrangement requirement used when this DataSource was created.

$sel:redshiftMetadata:DataSource', dataSource_redshiftMetadata - Undocumented member.

$sel:message:DataSource', dataSource_message - A description of the most recent details about creating the DataSource.

$sel:dataSourceId:DataSource', dataSource_dataSourceId - The ID that is assigned to the DataSource during creation.

$sel:computeStatistics:DataSource', dataSource_computeStatistics - The parameter is true if statistics need to be generated from the observation data.

$sel:createdAt:DataSource', dataSource_createdAt - The time that the DataSource was created. The time is expressed in epoch time.

$sel:dataLocationS3:DataSource', dataSource_dataLocationS3 - The location and name of the data in Amazon Simple Storage Service (Amazon S3) that is used by a DataSource.

$sel:name:DataSource', dataSource_name - A user-supplied name or description of the DataSource.

$sel:numberOfFiles:DataSource', dataSource_numberOfFiles - The number of data files referenced by the DataSource.

$sel:createdByIamUser:DataSource', dataSource_createdByIamUser - The AWS user account from which the DataSource was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

$sel:finishedAt:DataSource', dataSource_finishedAt - Undocumented member.

$sel:dataSizeInBytes:DataSource', dataSource_dataSizeInBytes - The total number of observations contained in the data files that the DataSource references.

$sel:computeTime:DataSource', dataSource_computeTime - Undocumented member.

$sel:rDSMetadata:DataSource', dataSource_rDSMetadata - Undocumented member.

$sel:lastUpdatedAt:DataSource', dataSource_lastUpdatedAt - The time of the most recent edit to the BatchPrediction. The time is expressed in epoch time.

dataSource_status :: Lens' DataSource (Maybe EntityStatus)

The current status of the DataSource. This element can have one of the following values:

PENDING - Amazon Machine Learning (Amazon ML) submitted a request to create a DataSource.

INPROGRESS - The creation process is underway.

FAILED - The request to create a DataSource did not run to completion. It is not usable.

COMPLETED - The creation process completed successfully.

DELETED - The DataSource is marked as deleted. It is not usable.

dataSource_startedAt :: Lens' DataSource (Maybe UTCTime)

Undocumented member.

dataSource_roleARN :: Lens' DataSource (Maybe Text)

Undocumented member.

dataSource_dataRearrangement :: Lens' DataSource (Maybe Text)

A JSON string that represents the splitting and rearrangement requirement used when this DataSource was created.

dataSource_redshiftMetadata :: Lens' DataSource (Maybe RedshiftMetadata)

Undocumented member.

dataSource_message :: Lens' DataSource (Maybe Text)

A description of the most recent details about creating the DataSource.

dataSource_dataSourceId :: Lens' DataSource (Maybe Text)

The ID that is assigned to the DataSource during creation.

dataSource_computeStatistics :: Lens' DataSource (Maybe Bool)

The parameter is true if statistics need to be generated from the observation data.

dataSource_createdAt :: Lens' DataSource (Maybe UTCTime)

The time that the DataSource was created. The time is expressed in epoch time.

dataSource_dataLocationS3 :: Lens' DataSource (Maybe Text)

The location and name of the data in Amazon Simple Storage Service (Amazon S3) that is used by a DataSource.

dataSource_name :: Lens' DataSource (Maybe Text)

A user-supplied name or description of the DataSource.

dataSource_numberOfFiles :: Lens' DataSource (Maybe Integer)

The number of data files referenced by the DataSource.

dataSource_createdByIamUser :: Lens' DataSource (Maybe Text)

The AWS user account from which the DataSource was created. The account type can be either an AWS root account or an AWS Identity and Access Management (IAM) user account.

dataSource_finishedAt :: Lens' DataSource (Maybe UTCTime)

Undocumented member.

dataSource_dataSizeInBytes :: Lens' DataSource (Maybe Integer)

The total number of observations contained in the data files that the DataSource references.

dataSource_computeTime :: Lens' DataSource (Maybe Integer)

Undocumented member.

dataSource_rDSMetadata :: Lens' DataSource (Maybe RDSMetadata)

Undocumented member.

dataSource_lastUpdatedAt :: Lens' DataSource (Maybe UTCTime)

The time of the most recent edit to the BatchPrediction. The time is expressed in epoch time.