Network.AWS.MachineLearning.UpdateDataSource

Description

Updates the DataSourceName of a DataSource.

You can use the GetDataSource operation to view the contents of the updated data element.

Creating a Request

data UpdateDataSource

See: newUpdateDataSource smart constructor.

Constructors

UpdateDataSource'

Fields

dataSourceId :: Text

The ID assigned to the DataSource during creation.

dataSourceName :: Text

A new user-supplied name or description of the DataSource that will replace the current description.

newUpdateDataSource

Arguments

:: Text
$sel:dataSourceId:UpdateDataSource'

-> Text
$sel:dataSourceName:UpdateDataSource'

-> UpdateDataSource

Create a value of UpdateDataSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataSourceId:UpdateDataSource', updateDataSource_dataSourceId - The ID assigned to the DataSource during creation.

$sel:dataSourceName:UpdateDataSource', updateDataSource_dataSourceName - A new user-supplied name or description of the DataSource that will replace the current description.

Request Lenses

updateDataSource_dataSourceId :: Lens' UpdateDataSource Text

The ID assigned to the DataSource during creation.

updateDataSource_dataSourceName :: Lens' UpdateDataSource Text

A new user-supplied name or description of the DataSource that will replace the current description.

Destructuring the Response

data UpdateDataSourceResponse

Represents the output of an UpdateDataSource operation.

You can see the updated content by using the GetBatchPrediction operation.

See: newUpdateDataSourceResponse smart constructor.

Constructors

UpdateDataSourceResponse'

Fields

dataSourceId :: Maybe Text

The ID assigned to the DataSource during creation. This value should be identical to the value of the DataSourceID in the request.

httpStatus :: Int

The response's http status code.

newUpdateDataSourceResponse

Arguments

:: Int
$sel:httpStatus:UpdateDataSourceResponse'

-> UpdateDataSourceResponse

Create a value of UpdateDataSourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataSourceId:UpdateDataSource', updateDataSourceResponse_dataSourceId - The ID assigned to the DataSource during creation. This value should be identical to the value of the DataSourceID in the request.

$sel:httpStatus:UpdateDataSourceResponse', updateDataSourceResponse_httpStatus - The response's http status code.

Response Lenses

updateDataSourceResponse_dataSourceId :: Lens' UpdateDataSourceResponse (Maybe Text)

The ID assigned to the DataSource during creation. This value should be identical to the value of the DataSourceID in the request.

updateDataSourceResponse_httpStatus :: Lens' UpdateDataSourceResponse Int

The response's http status code.