Network.AWS.Macie

Description

Derived from API version 2017-12-19 of the AWS service descriptions, licensed under Apache 2.0.

Amazon Macie Classic

Amazon Macie Classic is a security service that uses machine learning to automatically discover, classify, and protect sensitive data in AWS. Macie Classic recognizes sensitive data such as personally identifiable information (PII) or intellectual property, and provides you with dashboards and alerts that give visibility into how this data is being accessed or moved. For more information, see the Amazon Macie Classic User Guide.

Service Configuration

defaultService :: Service

API version 2017-12-19 of the Amazon Macie SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by Macie.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have required permissions to access the requested resource.

InvalidInputException

_InvalidInputException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because an invalid or out-of-range value was supplied for an input parameter.

InternalException

_InternalException :: AsError a => Getting (First ServiceError) a ServiceError

Internal server error.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because it attempted to create resources beyond the current AWS account limits. The error code describes the limit exceeded.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

AssociateS3Resources

data AssociateS3Resources

See: newAssociateS3Resources smart constructor.

Constructors

AssociateS3Resources' (Maybe Text) [S3ResourceClassification]

newAssociateS3Resources :: AssociateS3Resources

Create a value of AssociateS3Resources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memberAccountId:AssociateS3Resources', associateS3Resources_memberAccountId - The ID of the Amazon Macie Classic member account whose resources you want to associate with Macie Classic.

$sel:s3Resources:AssociateS3Resources', associateS3Resources_s3Resources - The S3 resources that you want to associate with Amazon Macie Classic for monitoring and data classification.

data AssociateS3ResourcesResponse

See: newAssociateS3ResourcesResponse smart constructor.

Constructors

AssociateS3ResourcesResponse' (Maybe [FailedS3Resource]) Int

newAssociateS3ResourcesResponse

Arguments

:: Int
$sel:httpStatus:AssociateS3ResourcesResponse'

-> AssociateS3ResourcesResponse

Create a value of AssociateS3ResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedS3Resources:AssociateS3ResourcesResponse', associateS3ResourcesResponse_failedS3Resources - S3 resources that couldn't be associated with Amazon Macie Classic. An error code and an error message are provided for each failed item.

$sel:httpStatus:AssociateS3ResourcesResponse', associateS3ResourcesResponse_httpStatus - The response's http status code.

AssociateMemberAccount

data AssociateMemberAccount

See: newAssociateMemberAccount smart constructor.

Constructors

AssociateMemberAccount' Text

newAssociateMemberAccount

Arguments

:: Text
$sel:memberAccountId:AssociateMemberAccount'

-> AssociateMemberAccount

Create a value of AssociateMemberAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memberAccountId:AssociateMemberAccount', associateMemberAccount_memberAccountId - The ID of the AWS account that you want to associate with Amazon Macie Classic as a member account.

data AssociateMemberAccountResponse

See: newAssociateMemberAccountResponse smart constructor.

Constructors

AssociateMemberAccountResponse'

newAssociateMemberAccountResponse :: AssociateMemberAccountResponse

Create a value of AssociateMemberAccountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

UpdateS3Resources

data UpdateS3Resources

See: newUpdateS3Resources smart constructor.

Constructors

UpdateS3Resources' (Maybe Text) [S3ResourceClassificationUpdate]

newUpdateS3Resources :: UpdateS3Resources

Create a value of UpdateS3Resources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memberAccountId:UpdateS3Resources', updateS3Resources_memberAccountId - The AWS ID of the Amazon Macie Classic member account whose S3 resources' classification types you want to update.

$sel:s3ResourcesUpdate:UpdateS3Resources', updateS3Resources_s3ResourcesUpdate - The S3 resources whose classification types you want to update.

data UpdateS3ResourcesResponse

See: newUpdateS3ResourcesResponse smart constructor.

Constructors

UpdateS3ResourcesResponse' (Maybe [FailedS3Resource]) Int

newUpdateS3ResourcesResponse

Arguments

:: Int
$sel:httpStatus:UpdateS3ResourcesResponse'

-> UpdateS3ResourcesResponse

Create a value of UpdateS3ResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedS3Resources:UpdateS3ResourcesResponse', updateS3ResourcesResponse_failedS3Resources - The S3 resources whose classification types can't be updated. An error code and an error message are provided for each failed item.

$sel:httpStatus:UpdateS3ResourcesResponse', updateS3ResourcesResponse_httpStatus - The response's http status code.

ListMemberAccounts (Paginated)

data ListMemberAccounts

See: newListMemberAccounts smart constructor.

Constructors

ListMemberAccounts' (Maybe Text) (Maybe Int)

newListMemberAccounts :: ListMemberAccounts

Create a value of ListMemberAccounts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListMemberAccounts', listMemberAccounts_nextToken - Use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListMemberAccounts action. Subsequent calls to the action fill nextToken in the request with the value of nextToken from the previous response to continue listing data.

$sel:maxResults:ListMemberAccounts', listMemberAccounts_maxResults - Use this parameter to indicate the maximum number of items that you want in the response. The default value is 250.

data ListMemberAccountsResponse

See: newListMemberAccountsResponse smart constructor.

Constructors

ListMemberAccountsResponse' (Maybe Text) (Maybe [MemberAccount]) Int

newListMemberAccountsResponse

Arguments

:: Int
$sel:httpStatus:ListMemberAccountsResponse'

-> ListMemberAccountsResponse

Create a value of ListMemberAccountsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListMemberAccounts', listMemberAccountsResponse_nextToken - When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

$sel:memberAccounts:ListMemberAccountsResponse', listMemberAccountsResponse_memberAccounts - A list of the Amazon Macie Classic member accounts returned by the action. The current Macie Classic administrator account is also included in this list.

$sel:httpStatus:ListMemberAccountsResponse', listMemberAccountsResponse_httpStatus - The response's http status code.

DisassociateMemberAccount

data DisassociateMemberAccount

See: newDisassociateMemberAccount smart constructor.

Constructors

DisassociateMemberAccount' Text

newDisassociateMemberAccount

Arguments

:: Text
$sel:memberAccountId:DisassociateMemberAccount'

-> DisassociateMemberAccount

Create a value of DisassociateMemberAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memberAccountId:DisassociateMemberAccount', disassociateMemberAccount_memberAccountId - The ID of the member account that you want to remove from Amazon Macie Classic.

data DisassociateMemberAccountResponse

See: newDisassociateMemberAccountResponse smart constructor.

Constructors

DisassociateMemberAccountResponse'

newDisassociateMemberAccountResponse :: DisassociateMemberAccountResponse

Create a value of DisassociateMemberAccountResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

ListS3Resources (Paginated)

data ListS3Resources

See: newListS3Resources smart constructor.

Constructors

ListS3Resources' (Maybe Text) (Maybe Text) (Maybe Int)

newListS3Resources :: ListS3Resources

Create a value of ListS3Resources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memberAccountId:ListS3Resources', listS3Resources_memberAccountId - The Amazon Macie Classic member account ID whose associated S3 resources you want to list.

$sel:nextToken:ListS3Resources', listS3Resources_nextToken - Use this parameter when paginating results. Set its value to null on your first call to the ListS3Resources action. Subsequent calls to the action fill nextToken in the request with the value of nextToken from the previous response to continue listing data.

$sel:maxResults:ListS3Resources', listS3Resources_maxResults - Use this parameter to indicate the maximum number of items that you want in the response. The default value is 250.

data ListS3ResourcesResponse

See: newListS3ResourcesResponse smart constructor.

Constructors

ListS3ResourcesResponse' (Maybe Text) (Maybe [S3ResourceClassification]) Int

newListS3ResourcesResponse

Arguments

:: Int
$sel:httpStatus:ListS3ResourcesResponse'

-> ListS3ResourcesResponse

Create a value of ListS3ResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListS3Resources', listS3ResourcesResponse_nextToken - When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

$sel:s3Resources:ListS3ResourcesResponse', listS3ResourcesResponse_s3Resources - A list of the associated S3 resources returned by the action.

$sel:httpStatus:ListS3ResourcesResponse', listS3ResourcesResponse_httpStatus - The response's http status code.

DisassociateS3Resources

data DisassociateS3Resources

See: newDisassociateS3Resources smart constructor.

Constructors

DisassociateS3Resources' (Maybe Text) [S3Resource]

newDisassociateS3Resources :: DisassociateS3Resources

Create a value of DisassociateS3Resources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memberAccountId:DisassociateS3Resources', disassociateS3Resources_memberAccountId - The ID of the Amazon Macie Classic member account whose resources you want to remove from being monitored by Macie Classic.

$sel:associatedS3Resources:DisassociateS3Resources', disassociateS3Resources_associatedS3Resources - The S3 resources (buckets or prefixes) that you want to remove from being monitored and classified by Amazon Macie Classic.

data DisassociateS3ResourcesResponse

See: newDisassociateS3ResourcesResponse smart constructor.

Constructors

DisassociateS3ResourcesResponse' (Maybe [FailedS3Resource]) Int

newDisassociateS3ResourcesResponse

Arguments

:: Int
$sel:httpStatus:DisassociateS3ResourcesResponse'

-> DisassociateS3ResourcesResponse

Create a value of DisassociateS3ResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedS3Resources:DisassociateS3ResourcesResponse', disassociateS3ResourcesResponse_failedS3Resources - S3 resources that couldn't be removed from being monitored and classified by Amazon Macie Classic. An error code and an error message are provided for each failed item.

$sel:httpStatus:DisassociateS3ResourcesResponse', disassociateS3ResourcesResponse_httpStatus - The response's http status code.

Types

S3ContinuousClassificationType

newtype S3ContinuousClassificationType

Constructors

S3ContinuousClassificationType'

Fields

fromS3ContinuousClassificationType :: Text

Bundled Patterns

pattern S3ContinuousClassificationType_FULL :: S3ContinuousClassificationType

S3OneTimeClassificationType

newtype S3OneTimeClassificationType

Constructors

S3OneTimeClassificationType'

Fields

fromS3OneTimeClassificationType :: Text

Bundled Patterns

pattern S3OneTimeClassificationType_FULL :: S3OneTimeClassificationType

pattern S3OneTimeClassificationType_NONE :: S3OneTimeClassificationType

ClassificationType

data ClassificationType

The classification type that Amazon Macie Classic applies to the associated S3 resources.

See: newClassificationType smart constructor.

Constructors

ClassificationType' S3OneTimeClassificationType S3ContinuousClassificationType

newClassificationType

Arguments

:: S3OneTimeClassificationType
$sel:oneTime:ClassificationType'

-> S3ContinuousClassificationType
$sel:continuous:ClassificationType'

-> ClassificationType

Create a value of ClassificationType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:oneTime:ClassificationType', classificationType_oneTime - A one-time classification of all of the existing objects in a specified S3 bucket.

$sel:continuous:ClassificationType', classificationType_continuous - A continuous classification of the objects that are added to a specified S3 bucket. Amazon Macie Classic begins performing continuous classification after a bucket is successfully associated with Macie Classic.

ClassificationTypeUpdate

data ClassificationTypeUpdate

The classification type that Amazon Macie Classic applies to the associated S3 resources. At least one of the classification types (oneTime or continuous) must be specified.

See: newClassificationTypeUpdate smart constructor.

Constructors

ClassificationTypeUpdate' (Maybe S3ContinuousClassificationType) (Maybe S3OneTimeClassificationType)

newClassificationTypeUpdate :: ClassificationTypeUpdate

Create a value of ClassificationTypeUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:continuous:ClassificationTypeUpdate', classificationTypeUpdate_continuous - A continuous classification of the objects that are added to a specified S3 bucket. Amazon Macie Classic begins performing continuous classification after a bucket is successfully associated with Macie Classic.

$sel:oneTime:ClassificationTypeUpdate', classificationTypeUpdate_oneTime - A one-time classification of all of the existing objects in a specified S3 bucket.

FailedS3Resource

data FailedS3Resource

Includes details about the failed S3 resources.

See: newFailedS3Resource smart constructor.

Constructors

FailedS3Resource' (Maybe Text) (Maybe Text) (Maybe S3Resource)

newFailedS3Resource :: FailedS3Resource

Create a value of FailedS3Resource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:errorCode:FailedS3Resource', failedS3Resource_errorCode - The status code of a failed item.

$sel:errorMessage:FailedS3Resource', failedS3Resource_errorMessage - The error message of a failed item.

$sel:failedItem:FailedS3Resource', failedS3Resource_failedItem - The failed S3 resources.

MemberAccount

data MemberAccount

Contains information about the Amazon Macie Classic member account.

See: newMemberAccount smart constructor.

Constructors

MemberAccount' (Maybe Text)

newMemberAccount :: MemberAccount

Create a value of MemberAccount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:MemberAccount', memberAccount_accountId - The AWS account ID of the Amazon Macie Classic member account.

S3Resource

data S3Resource

Contains information about the S3 resource. This data type is used as a request parameter in the DisassociateS3Resources action and can be used as a response parameter in the AssociateS3Resources and UpdateS3Resources actions.

See: newS3Resource smart constructor.

Constructors

S3Resource' (Maybe Text) Text

newS3Resource

Arguments

:: Text
$sel:bucketName:S3Resource'

-> S3Resource

Create a value of S3Resource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefix:S3Resource', s3Resource_prefix - The prefix of the S3 bucket.

$sel:bucketName:S3Resource', s3Resource_bucketName - The name of the S3 bucket.

S3ResourceClassification

data S3ResourceClassification

The S3 resources that you want to associate with Amazon Macie Classic for monitoring and data classification. This data type is used as a request parameter in the AssociateS3Resources action and a response parameter in the ListS3Resources action.

See: newS3ResourceClassification smart constructor.

Constructors

S3ResourceClassification' (Maybe Text) Text ClassificationType

newS3ResourceClassification

Arguments

:: Text
$sel:bucketName:S3ResourceClassification'

-> ClassificationType
$sel:classificationType:S3ResourceClassification'

-> S3ResourceClassification

Create a value of S3ResourceClassification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefix:S3ResourceClassification', s3ResourceClassification_prefix - The prefix of the S3 bucket that you want to associate with Amazon Macie Classic.

$sel:bucketName:S3ResourceClassification', s3ResourceClassification_bucketName - The name of the S3 bucket that you want to associate with Amazon Macie Classic.

$sel:classificationType:S3ResourceClassification', s3ResourceClassification_classificationType - The classification type that you want to specify for the resource associated with Amazon Macie Classic.

S3ResourceClassificationUpdate

data S3ResourceClassificationUpdate

The S3 resources whose classification types you want to update. This data type is used as a request parameter in the UpdateS3Resources action.

See: newS3ResourceClassificationUpdate smart constructor.

Constructors

S3ResourceClassificationUpdate' (Maybe Text) Text ClassificationTypeUpdate

newS3ResourceClassificationUpdate

Arguments

:: Text
$sel:bucketName:S3ResourceClassificationUpdate'

-> ClassificationTypeUpdate
$sel:classificationTypeUpdate:S3ResourceClassificationUpdate'

-> S3ResourceClassificationUpdate

Create a value of S3ResourceClassificationUpdate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:prefix:S3ResourceClassificationUpdate', s3ResourceClassificationUpdate_prefix - The prefix of the S3 bucket whose classification types you want to update.

$sel:bucketName:S3ResourceClassificationUpdate', s3ResourceClassificationUpdate_bucketName - The name of the S3 bucket whose classification types you want to update.

$sel:classificationTypeUpdate:S3ResourceClassificationUpdate', s3ResourceClassificationUpdate_classificationTypeUpdate - The classification type that you want to update for the resource associated with Amazon Macie Classic.