Network.AWS.Macie.AssociateS3Resources

Description

Associates specified S3 resources with Amazon Macie Classic for monitoring and data classification. If memberAccountId isn't specified, the action associates specified S3 resources with Macie Classic for the current Macie Classic administrator account. If memberAccountId is specified, the action associates specified S3 resources with Macie Classic for the specified member account.

Creating a Request

data AssociateS3Resources

See: newAssociateS3Resources smart constructor.

Constructors

AssociateS3Resources'

Fields

memberAccountId :: Maybe Text

The ID of the Amazon Macie Classic member account whose resources you want to associate with Macie Classic.

s3Resources :: [S3ResourceClassification]

The S3 resources that you want to associate with Amazon Macie Classic for monitoring and data classification.

newAssociateS3Resources :: AssociateS3Resources

Create a value of AssociateS3Resources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memberAccountId:AssociateS3Resources', associateS3Resources_memberAccountId - The ID of the Amazon Macie Classic member account whose resources you want to associate with Macie Classic.

$sel:s3Resources:AssociateS3Resources', associateS3Resources_s3Resources - The S3 resources that you want to associate with Amazon Macie Classic for monitoring and data classification.

Request Lenses

associateS3Resources_memberAccountId :: Lens' AssociateS3Resources (Maybe Text)

The ID of the Amazon Macie Classic member account whose resources you want to associate with Macie Classic.

associateS3Resources_s3Resources :: Lens' AssociateS3Resources [S3ResourceClassification]

The S3 resources that you want to associate with Amazon Macie Classic for monitoring and data classification.

Destructuring the Response

data AssociateS3ResourcesResponse

See: newAssociateS3ResourcesResponse smart constructor.

Constructors

AssociateS3ResourcesResponse'

Fields

failedS3Resources :: Maybe [FailedS3Resource]

S3 resources that couldn't be associated with Amazon Macie Classic. An error code and an error message are provided for each failed item.

httpStatus :: Int

The response's http status code.

newAssociateS3ResourcesResponse

Arguments

:: Int
$sel:httpStatus:AssociateS3ResourcesResponse'

-> AssociateS3ResourcesResponse

Create a value of AssociateS3ResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failedS3Resources:AssociateS3ResourcesResponse', associateS3ResourcesResponse_failedS3Resources - S3 resources that couldn't be associated with Amazon Macie Classic. An error code and an error message are provided for each failed item.

$sel:httpStatus:AssociateS3ResourcesResponse', associateS3ResourcesResponse_httpStatus - The response's http status code.

Response Lenses

associateS3ResourcesResponse_failedS3Resources :: Lens' AssociateS3ResourcesResponse (Maybe [FailedS3Resource])

S3 resources that couldn't be associated with Amazon Macie Classic. An error code and an error message are provided for each failed item.

associateS3ResourcesResponse_httpStatus :: Lens' AssociateS3ResourcesResponse Int

The response's http status code.