Network.AWS.MacieV2.CreateSampleFindings

Description

Creates sample findings.

Creating a Request

data CreateSampleFindings

See: newCreateSampleFindings smart constructor.

Constructors

CreateSampleFindings'

Fields

findingTypes :: Maybe [FindingType]

An array that lists one or more types of findings to include in the set of sample findings. Currently, the only supported value is Policy:IAMUser/S3BucketEncryptionDisabled.

newCreateSampleFindings :: CreateSampleFindings

Create a value of CreateSampleFindings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:findingTypes:CreateSampleFindings', createSampleFindings_findingTypes - An array that lists one or more types of findings to include in the set of sample findings. Currently, the only supported value is Policy:IAMUser/S3BucketEncryptionDisabled.

Request Lenses

createSampleFindings_findingTypes :: Lens' CreateSampleFindings (Maybe [FindingType])

An array that lists one or more types of findings to include in the set of sample findings. Currently, the only supported value is Policy:IAMUser/S3BucketEncryptionDisabled.

Destructuring the Response

data CreateSampleFindingsResponse

See: newCreateSampleFindingsResponse smart constructor.

Constructors

CreateSampleFindingsResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateSampleFindingsResponse

Arguments

:: Int
$sel:httpStatus:CreateSampleFindingsResponse'

-> CreateSampleFindingsResponse

Create a value of CreateSampleFindingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateSampleFindingsResponse', createSampleFindingsResponse_httpStatus - The response's http status code.

Response Lenses

createSampleFindingsResponse_httpStatus :: Lens' CreateSampleFindingsResponse Int

The response's http status code.