Network.AWS.MacieV2.GetCustomDataIdentifier

Description

Retrieves the criteria and other settings for a custom data identifier.

Creating a Request

data GetCustomDataIdentifier

See: newGetCustomDataIdentifier smart constructor.

Constructors

GetCustomDataIdentifier'

Fields

id :: Text

The unique identifier for the Amazon Macie resource or account that the request applies to.

newGetCustomDataIdentifier

Arguments

:: Text
$sel:id:GetCustomDataIdentifier'

-> GetCustomDataIdentifier

Create a value of GetCustomDataIdentifier with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:GetCustomDataIdentifier', getCustomDataIdentifier_id - The unique identifier for the Amazon Macie resource or account that the request applies to.

Request Lenses

getCustomDataIdentifier_id :: Lens' GetCustomDataIdentifier Text

The unique identifier for the Amazon Macie resource or account that the request applies to.

Destructuring the Response

data GetCustomDataIdentifierResponse

See: newGetCustomDataIdentifierResponse smart constructor.

Constructors

GetCustomDataIdentifierResponse'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the custom data identifier.

createdAt :: Maybe POSIX

The date and time, in UTC and extended ISO 8601 format, when the custom data identifier was created.

regex :: Maybe Text

The regular expression (regex) that defines the pattern to match.

name :: Maybe Text

The custom name of the custom data identifier.

keywords :: Maybe [Text]

An array that lists specific character sequences (keywords), one of which must be within proximity (maximumMatchDistance) of the regular expression to match. Keywords aren't case sensitive.

ignoreWords :: Maybe [Text]

An array that lists specific character sequences (ignore words) to exclude from the results. If the text matched by the regular expression is the same as any string in this array, Amazon Macie ignores it. Ignore words are case sensitive.

id :: Maybe Text

The unique identifier for the custom data identifier.

deleted :: Maybe Bool

Specifies whether the custom data identifier was deleted. If you delete a custom data identifier, Amazon Macie doesn't delete it permanently. Instead, it soft deletes the identifier.

maximumMatchDistance :: Maybe Int

The maximum number of characters that can exist between text that matches the regex pattern and the character sequences specified by the keywords array. Amazon Macie includes or excludes a result based on the proximity of a keyword to text that matches the regex pattern.

description :: Maybe Text

The custom description of the custom data identifier.

tags :: Maybe (HashMap Text Text)

A map of key-value pairs that identifies the tags (keys and values) that are associated with the custom data identifier.

httpStatus :: Int

The response's http status code.

newGetCustomDataIdentifierResponse

Arguments

:: Int
$sel:httpStatus:GetCustomDataIdentifierResponse'

-> GetCustomDataIdentifierResponse

Create a value of GetCustomDataIdentifierResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:GetCustomDataIdentifierResponse', getCustomDataIdentifierResponse_arn - The Amazon Resource Name (ARN) of the custom data identifier.

$sel:createdAt:GetCustomDataIdentifierResponse', getCustomDataIdentifierResponse_createdAt - The date and time, in UTC and extended ISO 8601 format, when the custom data identifier was created.

$sel:regex:GetCustomDataIdentifierResponse', getCustomDataIdentifierResponse_regex - The regular expression (regex) that defines the pattern to match.

$sel:name:GetCustomDataIdentifierResponse', getCustomDataIdentifierResponse_name - The custom name of the custom data identifier.

$sel:keywords:GetCustomDataIdentifierResponse', getCustomDataIdentifierResponse_keywords - An array that lists specific character sequences (keywords), one of which must be within proximity (maximumMatchDistance) of the regular expression to match. Keywords aren't case sensitive.

$sel:ignoreWords:GetCustomDataIdentifierResponse', getCustomDataIdentifierResponse_ignoreWords - An array that lists specific character sequences (ignore words) to exclude from the results. If the text matched by the regular expression is the same as any string in this array, Amazon Macie ignores it. Ignore words are case sensitive.

$sel:id:GetCustomDataIdentifier', getCustomDataIdentifierResponse_id - The unique identifier for the custom data identifier.

$sel:deleted:GetCustomDataIdentifierResponse', getCustomDataIdentifierResponse_deleted - Specifies whether the custom data identifier was deleted. If you delete a custom data identifier, Amazon Macie doesn't delete it permanently. Instead, it soft deletes the identifier.

$sel:maximumMatchDistance:GetCustomDataIdentifierResponse', getCustomDataIdentifierResponse_maximumMatchDistance - The maximum number of characters that can exist between text that matches the regex pattern and the character sequences specified by the keywords array. Amazon Macie includes or excludes a result based on the proximity of a keyword to text that matches the regex pattern.

$sel:description:GetCustomDataIdentifierResponse', getCustomDataIdentifierResponse_description - The custom description of the custom data identifier.

$sel:tags:GetCustomDataIdentifierResponse', getCustomDataIdentifierResponse_tags - A map of key-value pairs that identifies the tags (keys and values) that are associated with the custom data identifier.

$sel:httpStatus:GetCustomDataIdentifierResponse', getCustomDataIdentifierResponse_httpStatus - The response's http status code.

Response Lenses

getCustomDataIdentifierResponse_arn :: Lens' GetCustomDataIdentifierResponse (Maybe Text)

The Amazon Resource Name (ARN) of the custom data identifier.

getCustomDataIdentifierResponse_createdAt :: Lens' GetCustomDataIdentifierResponse (Maybe UTCTime)

The date and time, in UTC and extended ISO 8601 format, when the custom data identifier was created.

getCustomDataIdentifierResponse_regex :: Lens' GetCustomDataIdentifierResponse (Maybe Text)

The regular expression (regex) that defines the pattern to match.

getCustomDataIdentifierResponse_name :: Lens' GetCustomDataIdentifierResponse (Maybe Text)

The custom name of the custom data identifier.

getCustomDataIdentifierResponse_keywords :: Lens' GetCustomDataIdentifierResponse (Maybe [Text])

An array that lists specific character sequences (keywords), one of which must be within proximity (maximumMatchDistance) of the regular expression to match. Keywords aren't case sensitive.

getCustomDataIdentifierResponse_ignoreWords :: Lens' GetCustomDataIdentifierResponse (Maybe [Text])

An array that lists specific character sequences (ignore words) to exclude from the results. If the text matched by the regular expression is the same as any string in this array, Amazon Macie ignores it. Ignore words are case sensitive.

getCustomDataIdentifierResponse_id :: Lens' GetCustomDataIdentifierResponse (Maybe Text)

The unique identifier for the custom data identifier.

getCustomDataIdentifierResponse_deleted :: Lens' GetCustomDataIdentifierResponse (Maybe Bool)

Specifies whether the custom data identifier was deleted. If you delete a custom data identifier, Amazon Macie doesn't delete it permanently. Instead, it soft deletes the identifier.

getCustomDataIdentifierResponse_maximumMatchDistance :: Lens' GetCustomDataIdentifierResponse (Maybe Int)

The maximum number of characters that can exist between text that matches the regex pattern and the character sequences specified by the keywords array. Amazon Macie includes or excludes a result based on the proximity of a keyword to text that matches the regex pattern.

getCustomDataIdentifierResponse_description :: Lens' GetCustomDataIdentifierResponse (Maybe Text)

The custom description of the custom data identifier.

getCustomDataIdentifierResponse_tags :: Lens' GetCustomDataIdentifierResponse (Maybe (HashMap Text Text))

A map of key-value pairs that identifies the tags (keys and values) that are associated with the custom data identifier.

getCustomDataIdentifierResponse_httpStatus :: Lens' GetCustomDataIdentifierResponse Int

The response's http status code.