Network.AWS.MacieV2.GetFindingStatistics

Description

Retrieves (queries) aggregated statistical data about findings.

Creating a Request

data GetFindingStatistics

See: newGetFindingStatistics smart constructor.

Constructors

GetFindingStatistics'

Fields

size :: Maybe Int

The maximum number of items to include in each page of the response.

findingCriteria :: Maybe FindingCriteria

The criteria to use to filter the query results.

sortCriteria :: Maybe FindingStatisticsSortCriteria

The criteria to use to sort the query results.

groupBy :: GroupBy

The finding property to use to group the query results. Valid values are:

classificationDetails.jobId - The unique identifier for the classification job that produced the finding.

resourcesAffected.s3Bucket.name - The name of the S3 bucket that the finding applies to.

severity.description - The severity level of the finding, such as High or Medium.

type - The type of finding, such as Policy:IAMUser/S3BucketPublic and SensitiveData:S3Object/Personal.

newGetFindingStatistics

Arguments

:: GroupBy
$sel:groupBy:GetFindingStatistics'

-> GetFindingStatistics

Create a value of GetFindingStatistics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:size:GetFindingStatistics', getFindingStatistics_size - The maximum number of items to include in each page of the response.

$sel:findingCriteria:GetFindingStatistics', getFindingStatistics_findingCriteria - The criteria to use to filter the query results.

$sel:sortCriteria:GetFindingStatistics', getFindingStatistics_sortCriteria - The criteria to use to sort the query results.

$sel:groupBy:GetFindingStatistics', getFindingStatistics_groupBy - The finding property to use to group the query results. Valid values are:

classificationDetails.jobId - The unique identifier for the classification job that produced the finding.

resourcesAffected.s3Bucket.name - The name of the S3 bucket that the finding applies to.

severity.description - The severity level of the finding, such as High or Medium.

type - The type of finding, such as Policy:IAMUser/S3BucketPublic and SensitiveData:S3Object/Personal.

Request Lenses

getFindingStatistics_size :: Lens' GetFindingStatistics (Maybe Int)

The maximum number of items to include in each page of the response.

getFindingStatistics_findingCriteria :: Lens' GetFindingStatistics (Maybe FindingCriteria)

The criteria to use to filter the query results.

getFindingStatistics_sortCriteria :: Lens' GetFindingStatistics (Maybe FindingStatisticsSortCriteria)

The criteria to use to sort the query results.

getFindingStatistics_groupBy :: Lens' GetFindingStatistics GroupBy

The finding property to use to group the query results. Valid values are:

classificationDetails.jobId - The unique identifier for the classification job that produced the finding.

resourcesAffected.s3Bucket.name - The name of the S3 bucket that the finding applies to.

severity.description - The severity level of the finding, such as High or Medium.

type - The type of finding, such as Policy:IAMUser/S3BucketPublic and SensitiveData:S3Object/Personal.

Destructuring the Response

data GetFindingStatisticsResponse

See: newGetFindingStatisticsResponse smart constructor.

Constructors

GetFindingStatisticsResponse'

Fields

countsByGroup :: Maybe [GroupCount]

An array of objects, one for each group of findings that meet the filter criteria specified in the request.

httpStatus :: Int

The response's http status code.

newGetFindingStatisticsResponse

Arguments

:: Int
$sel:httpStatus:GetFindingStatisticsResponse'

-> GetFindingStatisticsResponse

Create a value of GetFindingStatisticsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:countsByGroup:GetFindingStatisticsResponse', getFindingStatisticsResponse_countsByGroup - An array of objects, one for each group of findings that meet the filter criteria specified in the request.

$sel:httpStatus:GetFindingStatisticsResponse', getFindingStatisticsResponse_httpStatus - The response's http status code.

Response Lenses

getFindingStatisticsResponse_countsByGroup :: Lens' GetFindingStatisticsResponse (Maybe [GroupCount])

An array of objects, one for each group of findings that meet the filter criteria specified in the request.

getFindingStatisticsResponse_httpStatus :: Lens' GetFindingStatisticsResponse Int

The response's http status code.