Network.AWS.MacieV2.GetMacieSession

Description

Retrieves the current status and configuration settings for an Amazon Macie account.

Creating a Request

data GetMacieSession

See: newGetMacieSession smart constructor.

Constructors

GetMacieSession'

newGetMacieSession :: GetMacieSession

Create a value of GetMacieSession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data GetMacieSessionResponse

See: newGetMacieSessionResponse smart constructor.

Constructors

GetMacieSessionResponse'

Fields

status :: Maybe MacieStatus

The current status of the Macie account. Possible values are: PAUSED, the account is enabled but all Macie activities are suspended (paused) for the account; and, ENABLED, the account is enabled and all Macie activities are enabled for the account.

createdAt :: Maybe POSIX

The date and time, in UTC and extended ISO 8601 format, when the Amazon Macie account was created.

findingPublishingFrequency :: Maybe FindingPublishingFrequency

The frequency with which Macie publishes updates to policy findings for the account. This includes publishing updates to Security Hub and Amazon EventBridge (formerly called Amazon CloudWatch Events).

updatedAt :: Maybe POSIX

The date and time, in UTC and extended ISO 8601 format, of the most recent change to the status of the Macie account.

serviceRole :: Maybe Text

The Amazon Resource Name (ARN) of the service-linked role that allows Macie to monitor and analyze data in Amazon Web Services resources for the account.

httpStatus :: Int

The response's http status code.

newGetMacieSessionResponse

Arguments

:: Int
$sel:httpStatus:GetMacieSessionResponse'

-> GetMacieSessionResponse

Create a value of GetMacieSessionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:GetMacieSessionResponse', getMacieSessionResponse_status - The current status of the Macie account. Possible values are: PAUSED, the account is enabled but all Macie activities are suspended (paused) for the account; and, ENABLED, the account is enabled and all Macie activities are enabled for the account.

$sel:createdAt:GetMacieSessionResponse', getMacieSessionResponse_createdAt - The date and time, in UTC and extended ISO 8601 format, when the Amazon Macie account was created.

$sel:findingPublishingFrequency:GetMacieSessionResponse', getMacieSessionResponse_findingPublishingFrequency - The frequency with which Macie publishes updates to policy findings for the account. This includes publishing updates to Security Hub and Amazon EventBridge (formerly called Amazon CloudWatch Events).

$sel:updatedAt:GetMacieSessionResponse', getMacieSessionResponse_updatedAt - The date and time, in UTC and extended ISO 8601 format, of the most recent change to the status of the Macie account.

$sel:serviceRole:GetMacieSessionResponse', getMacieSessionResponse_serviceRole - The Amazon Resource Name (ARN) of the service-linked role that allows Macie to monitor and analyze data in Amazon Web Services resources for the account.

$sel:httpStatus:GetMacieSessionResponse', getMacieSessionResponse_httpStatus - The response's http status code.

Response Lenses

getMacieSessionResponse_status :: Lens' GetMacieSessionResponse (Maybe MacieStatus)

The current status of the Macie account. Possible values are: PAUSED, the account is enabled but all Macie activities are suspended (paused) for the account; and, ENABLED, the account is enabled and all Macie activities are enabled for the account.

getMacieSessionResponse_createdAt :: Lens' GetMacieSessionResponse (Maybe UTCTime)

The date and time, in UTC and extended ISO 8601 format, when the Amazon Macie account was created.

getMacieSessionResponse_findingPublishingFrequency :: Lens' GetMacieSessionResponse (Maybe FindingPublishingFrequency)

The frequency with which Macie publishes updates to policy findings for the account. This includes publishing updates to Security Hub and Amazon EventBridge (formerly called Amazon CloudWatch Events).

getMacieSessionResponse_updatedAt :: Lens' GetMacieSessionResponse (Maybe UTCTime)

The date and time, in UTC and extended ISO 8601 format, of the most recent change to the status of the Macie account.

getMacieSessionResponse_serviceRole :: Lens' GetMacieSessionResponse (Maybe Text)

The Amazon Resource Name (ARN) of the service-linked role that allows Macie to monitor and analyze data in Amazon Web Services resources for the account.

getMacieSessionResponse_httpStatus :: Lens' GetMacieSessionResponse Int

The response's http status code.