Network.AWS.MacieV2.GetMember

Description

Retrieves information about an account that's associated with an Amazon Macie administrator account.

Creating a Request

data GetMember

See: newGetMember smart constructor.

Constructors

GetMember'

Fields

id :: Text

The unique identifier for the Amazon Macie resource or account that the request applies to.

newGetMember

Arguments

:: Text
$sel:id:GetMember'

-> GetMember

Create a value of GetMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:GetMember', getMember_id - The unique identifier for the Amazon Macie resource or account that the request applies to.

Request Lenses

getMember_id :: Lens' GetMember Text

The unique identifier for the Amazon Macie resource or account that the request applies to.

Destructuring the Response

data GetMemberResponse

See: newGetMemberResponse smart constructor.

Constructors

GetMemberResponse'

Fields

email :: Maybe Text

The email address for the account.

administratorAccountId :: Maybe Text

The Amazon Web Services account ID for the administrator account.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the account.

relationshipStatus :: Maybe RelationshipStatus

The current status of the relationship between the account and the administrator account.

masterAccountId :: Maybe Text

(Deprecated) The Amazon Web Services account ID for the administrator account. This property has been replaced by the administratorAccountId property and is retained only for backward compatibility.

invitedAt :: Maybe POSIX

The date and time, in UTC and extended ISO 8601 format, when an Amazon Macie membership invitation was last sent to the account. This value is null if a Macie invitation hasn't been sent to the account.

accountId :: Maybe Text

The Amazon Web Services account ID for the account.

updatedAt :: Maybe POSIX

The date and time, in UTC and extended ISO 8601 format, of the most recent change to the status of the relationship between the account and the administrator account.

tags :: Maybe (HashMap Text Text)

A map of key-value pairs that identifies the tags (keys and values) that are associated with the member account in Amazon Macie.

httpStatus :: Int

The response's http status code.

newGetMemberResponse

Arguments

:: Int
$sel:httpStatus:GetMemberResponse'

-> GetMemberResponse

Create a value of GetMemberResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:email:GetMemberResponse', getMemberResponse_email - The email address for the account.

$sel:administratorAccountId:GetMemberResponse', getMemberResponse_administratorAccountId - The Amazon Web Services account ID for the administrator account.

$sel:arn:GetMemberResponse', getMemberResponse_arn - The Amazon Resource Name (ARN) of the account.

$sel:relationshipStatus:GetMemberResponse', getMemberResponse_relationshipStatus - The current status of the relationship between the account and the administrator account.

$sel:masterAccountId:GetMemberResponse', getMemberResponse_masterAccountId - (Deprecated) The Amazon Web Services account ID for the administrator account. This property has been replaced by the administratorAccountId property and is retained only for backward compatibility.

$sel:invitedAt:GetMemberResponse', getMemberResponse_invitedAt - The date and time, in UTC and extended ISO 8601 format, when an Amazon Macie membership invitation was last sent to the account. This value is null if a Macie invitation hasn't been sent to the account.

$sel:accountId:GetMemberResponse', getMemberResponse_accountId - The Amazon Web Services account ID for the account.

$sel:updatedAt:GetMemberResponse', getMemberResponse_updatedAt - The date and time, in UTC and extended ISO 8601 format, of the most recent change to the status of the relationship between the account and the administrator account.

$sel:tags:GetMemberResponse', getMemberResponse_tags - A map of key-value pairs that identifies the tags (keys and values) that are associated with the member account in Amazon Macie.

$sel:httpStatus:GetMemberResponse', getMemberResponse_httpStatus - The response's http status code.

Response Lenses

getMemberResponse_email :: Lens' GetMemberResponse (Maybe Text)

The email address for the account.

getMemberResponse_administratorAccountId :: Lens' GetMemberResponse (Maybe Text)

The Amazon Web Services account ID for the administrator account.

getMemberResponse_arn :: Lens' GetMemberResponse (Maybe Text)

The Amazon Resource Name (ARN) of the account.

getMemberResponse_relationshipStatus :: Lens' GetMemberResponse (Maybe RelationshipStatus)

The current status of the relationship between the account and the administrator account.

getMemberResponse_masterAccountId :: Lens' GetMemberResponse (Maybe Text)

(Deprecated) The Amazon Web Services account ID for the administrator account. This property has been replaced by the administratorAccountId property and is retained only for backward compatibility.

getMemberResponse_invitedAt :: Lens' GetMemberResponse (Maybe UTCTime)

The date and time, in UTC and extended ISO 8601 format, when an Amazon Macie membership invitation was last sent to the account. This value is null if a Macie invitation hasn't been sent to the account.

getMemberResponse_accountId :: Lens' GetMemberResponse (Maybe Text)

The Amazon Web Services account ID for the account.

getMemberResponse_updatedAt :: Lens' GetMemberResponse (Maybe UTCTime)

The date and time, in UTC and extended ISO 8601 format, of the most recent change to the status of the relationship between the account and the administrator account.

getMemberResponse_tags :: Lens' GetMemberResponse (Maybe (HashMap Text Text))

A map of key-value pairs that identifies the tags (keys and values) that are associated with the member account in Amazon Macie.

getMemberResponse_httpStatus :: Lens' GetMemberResponse Int

The response's http status code.