Network.AWS.MacieV2.PutClassificationExportConfiguration

Description

Creates or updates the configuration settings for storing data classification results.

Creating a Request

data PutClassificationExportConfiguration

See: newPutClassificationExportConfiguration smart constructor.

Constructors

PutClassificationExportConfiguration'

Fields

configuration :: ClassificationExportConfiguration

The location to store data classification results in, and the encryption settings to use when storing results in that location.

newPutClassificationExportConfiguration

Arguments

:: ClassificationExportConfiguration
$sel:configuration:PutClassificationExportConfiguration'

-> PutClassificationExportConfiguration

Create a value of PutClassificationExportConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configuration:PutClassificationExportConfiguration', putClassificationExportConfiguration_configuration - The location to store data classification results in, and the encryption settings to use when storing results in that location.

Request Lenses

putClassificationExportConfiguration_configuration :: Lens' PutClassificationExportConfiguration ClassificationExportConfiguration

The location to store data classification results in, and the encryption settings to use when storing results in that location.

Destructuring the Response

data PutClassificationExportConfigurationResponse

See: newPutClassificationExportConfigurationResponse smart constructor.

Constructors

PutClassificationExportConfigurationResponse'

Fields

configuration :: Maybe ClassificationExportConfiguration

The location where the data classification results are stored, and the encryption settings that are used when storing results in that location.

httpStatus :: Int

The response's http status code.

newPutClassificationExportConfigurationResponse

Arguments

:: Int
$sel:httpStatus:PutClassificationExportConfigurationResponse'

-> PutClassificationExportConfigurationResponse

Create a value of PutClassificationExportConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configuration:PutClassificationExportConfiguration', putClassificationExportConfigurationResponse_configuration - The location where the data classification results are stored, and the encryption settings that are used when storing results in that location.

$sel:httpStatus:PutClassificationExportConfigurationResponse', putClassificationExportConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

putClassificationExportConfigurationResponse_configuration :: Lens' PutClassificationExportConfigurationResponse (Maybe ClassificationExportConfiguration)

The location where the data classification results are stored, and the encryption settings that are used when storing results in that location.

putClassificationExportConfigurationResponse_httpStatus :: Lens' PutClassificationExportConfigurationResponse Int

The response's http status code.