Network.AWS.MacieV2.SearchResources

Description

Retrieves (queries) statistical data and other information about Amazon Web Services resources that Amazon Macie monitors and analyzes.

This operation returns paginated results.

Creating a Request

data SearchResources

See: newSearchResources smart constructor.

Constructors

SearchResources'

Fields

bucketCriteria :: Maybe SearchResourcesBucketCriteria

The filter conditions that determine which S3 buckets to include or exclude from the query results.

sortCriteria :: Maybe SearchResourcesSortCriteria

The criteria to use to sort the results.

nextToken :: Maybe Text

The nextToken string that specifies which page of results to return in a paginated response.

maxResults :: Maybe Int

The maximum number of items to include in each page of the response. The default value is 50.

newSearchResources :: SearchResources

Create a value of SearchResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucketCriteria:SearchResources', searchResources_bucketCriteria - The filter conditions that determine which S3 buckets to include or exclude from the query results.

$sel:sortCriteria:SearchResources', searchResources_sortCriteria - The criteria to use to sort the results.

$sel:nextToken:SearchResources', searchResources_nextToken - The nextToken string that specifies which page of results to return in a paginated response.

$sel:maxResults:SearchResources', searchResources_maxResults - The maximum number of items to include in each page of the response. The default value is 50.

Request Lenses

searchResources_bucketCriteria :: Lens' SearchResources (Maybe SearchResourcesBucketCriteria)

The filter conditions that determine which S3 buckets to include or exclude from the query results.

searchResources_sortCriteria :: Lens' SearchResources (Maybe SearchResourcesSortCriteria)

The criteria to use to sort the results.

searchResources_nextToken :: Lens' SearchResources (Maybe Text)

The nextToken string that specifies which page of results to return in a paginated response.

searchResources_maxResults :: Lens' SearchResources (Maybe Int)

The maximum number of items to include in each page of the response. The default value is 50.

Destructuring the Response

data SearchResourcesResponse

See: newSearchResourcesResponse smart constructor.

Constructors

SearchResourcesResponse'

Fields

nextToken :: Maybe Text

The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

matchingResources :: Maybe [MatchingResource]

An array of objects, one for each resource that meets the filter criteria specified in the request.

httpStatus :: Int

The response's http status code.

newSearchResourcesResponse

Arguments

:: Int
$sel:httpStatus:SearchResourcesResponse'

-> SearchResourcesResponse

Create a value of SearchResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:SearchResources', searchResourcesResponse_nextToken - The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

$sel:matchingResources:SearchResourcesResponse', searchResourcesResponse_matchingResources - An array of objects, one for each resource that meets the filter criteria specified in the request.

$sel:httpStatus:SearchResourcesResponse', searchResourcesResponse_httpStatus - The response's http status code.

Response Lenses

searchResourcesResponse_nextToken :: Lens' SearchResourcesResponse (Maybe Text)

The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

searchResourcesResponse_matchingResources :: Lens' SearchResourcesResponse (Maybe [MatchingResource])

An array of objects, one for each resource that meets the filter criteria specified in the request.

searchResourcesResponse_httpStatus :: Lens' SearchResourcesResponse Int

The response's http status code.