Network.AWS.MacieV2.Types.BucketCriteriaAdditionalProperties

Description

Documentation

data BucketCriteriaAdditionalProperties

Specifies the operator to use in a property-based condition that filters the results of a query for information about S3 buckets.

See: newBucketCriteriaAdditionalProperties smart constructor.

Constructors

BucketCriteriaAdditionalProperties'

Fields

eq :: Maybe [Text]

The value for the property matches (equals) the specified value. If you specify multiple values, Amazon Macie uses OR logic to join the values.

lte :: Maybe Integer

The value for the property is less than or equal to the specified value.

prefix :: Maybe Text

The name of the bucket begins with the specified value.

gt :: Maybe Integer

The value for the property is greater than the specified value.

neq :: Maybe [Text]

The value for the property doesn't match (doesn't equal) the specified value. If you specify multiple values, Amazon Macie uses OR logic to join the values.

lt :: Maybe Integer

The value for the property is less than the specified value.

gte :: Maybe Integer

The value for the property is greater than or equal to the specified value.

newBucketCriteriaAdditionalProperties :: BucketCriteriaAdditionalProperties

Create a value of BucketCriteriaAdditionalProperties with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eq:BucketCriteriaAdditionalProperties', bucketCriteriaAdditionalProperties_eq - The value for the property matches (equals) the specified value. If you specify multiple values, Amazon Macie uses OR logic to join the values.

$sel:lte:BucketCriteriaAdditionalProperties', bucketCriteriaAdditionalProperties_lte - The value for the property is less than or equal to the specified value.

$sel:prefix:BucketCriteriaAdditionalProperties', bucketCriteriaAdditionalProperties_prefix - The name of the bucket begins with the specified value.

$sel:gt:BucketCriteriaAdditionalProperties', bucketCriteriaAdditionalProperties_gt - The value for the property is greater than the specified value.

$sel:neq:BucketCriteriaAdditionalProperties', bucketCriteriaAdditionalProperties_neq - The value for the property doesn't match (doesn't equal) the specified value. If you specify multiple values, Amazon Macie uses OR logic to join the values.

$sel:lt:BucketCriteriaAdditionalProperties', bucketCriteriaAdditionalProperties_lt - The value for the property is less than the specified value.

$sel:gte:BucketCriteriaAdditionalProperties', bucketCriteriaAdditionalProperties_gte - The value for the property is greater than or equal to the specified value.

bucketCriteriaAdditionalProperties_eq :: Lens' BucketCriteriaAdditionalProperties (Maybe [Text])

The value for the property matches (equals) the specified value. If you specify multiple values, Amazon Macie uses OR logic to join the values.

bucketCriteriaAdditionalProperties_lte :: Lens' BucketCriteriaAdditionalProperties (Maybe Integer)

The value for the property is less than or equal to the specified value.

bucketCriteriaAdditionalProperties_prefix :: Lens' BucketCriteriaAdditionalProperties (Maybe Text)

The name of the bucket begins with the specified value.

bucketCriteriaAdditionalProperties_gt :: Lens' BucketCriteriaAdditionalProperties (Maybe Integer)

The value for the property is greater than the specified value.

bucketCriteriaAdditionalProperties_neq :: Lens' BucketCriteriaAdditionalProperties (Maybe [Text])

The value for the property doesn't match (doesn't equal) the specified value. If you specify multiple values, Amazon Macie uses OR logic to join the values.

bucketCriteriaAdditionalProperties_lt :: Lens' BucketCriteriaAdditionalProperties (Maybe Integer)

The value for the property is less than the specified value.

bucketCriteriaAdditionalProperties_gte :: Lens' BucketCriteriaAdditionalProperties (Maybe Integer)

The value for the property is greater than or equal to the specified value.