Network.AWS.MacieV2.Types.BucketSortCriteria

Description

Documentation

data BucketSortCriteria

Specifies criteria for sorting the results of a query for information about S3 buckets.

See: newBucketSortCriteria smart constructor.

Constructors

BucketSortCriteria'

Fields

orderBy :: Maybe OrderBy

The sort order to apply to the results, based on the value specified by the attributeName property. Valid values are: ASC, sort the results in ascending order; and, DESC, sort the results in descending order.

attributeName :: Maybe Text

The name of the bucket property to sort the results by. This value can be one of the following properties that Amazon Macie defines as bucket metadata: accountId, bucketName, classifiableObjectCount, classifiableSizeInBytes, objectCount, or sizeInBytes.

newBucketSortCriteria :: BucketSortCriteria

Create a value of BucketSortCriteria with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:orderBy:BucketSortCriteria', bucketSortCriteria_orderBy - The sort order to apply to the results, based on the value specified by the attributeName property. Valid values are: ASC, sort the results in ascending order; and, DESC, sort the results in descending order.

$sel:attributeName:BucketSortCriteria', bucketSortCriteria_attributeName - The name of the bucket property to sort the results by. This value can be one of the following properties that Amazon Macie defines as bucket metadata: accountId, bucketName, classifiableObjectCount, classifiableSizeInBytes, objectCount, or sizeInBytes.

bucketSortCriteria_orderBy :: Lens' BucketSortCriteria (Maybe OrderBy)

The sort order to apply to the results, based on the value specified by the attributeName property. Valid values are: ASC, sort the results in ascending order; and, DESC, sort the results in descending order.

bucketSortCriteria_attributeName :: Lens' BucketSortCriteria (Maybe Text)

The name of the bucket property to sort the results by. This value can be one of the following properties that Amazon Macie defines as bucket metadata: accountId, bucketName, classifiableObjectCount, classifiableSizeInBytes, objectCount, or sizeInBytes.