Network.AWS.MacieV2.Types.CustomDataIdentifiers

Description

Documentation

data CustomDataIdentifiers

Provides information about custom data identifiers that produced a sensitive data finding, and the number of occurrences of the data that they detected for the finding.

See: newCustomDataIdentifiers smart constructor.

Constructors

CustomDataIdentifiers'

Fields

detections :: Maybe [CustomDetection]

The custom data identifiers that detected the data, and the number of occurrences of the data that each identifier detected.

totalCount :: Maybe Integer

The total number of occurrences of the data that was detected by the custom data identifiers and produced the finding.

newCustomDataIdentifiers :: CustomDataIdentifiers

Create a value of CustomDataIdentifiers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:detections:CustomDataIdentifiers', customDataIdentifiers_detections - The custom data identifiers that detected the data, and the number of occurrences of the data that each identifier detected.

$sel:totalCount:CustomDataIdentifiers', customDataIdentifiers_totalCount - The total number of occurrences of the data that was detected by the custom data identifiers and produced the finding.

customDataIdentifiers_detections :: Lens' CustomDataIdentifiers (Maybe [CustomDetection])

The custom data identifiers that detected the data, and the number of occurrences of the data that each identifier detected.

customDataIdentifiers_totalCount :: Lens' CustomDataIdentifiers (Maybe Integer)

The total number of occurrences of the data that was detected by the custom data identifiers and produced the finding.