Network.AWS.MacieV2.Types.CustomDetection

Description

Documentation

data CustomDetection

Provides information about a custom data identifier that produced a sensitive data finding, and the sensitive data that it detected for the finding.

See: newCustomDetection smart constructor.

Constructors

CustomDetection'

Fields

occurrences :: Maybe Occurrences

The location of 1-15 occurrences of the sensitive data that the custom data identifier detected. A finding includes location data for a maximum of 15 occurrences of sensitive data.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the custom data identifier.

count :: Maybe Integer

The total number of occurrences of the sensitive data that the custom data identifier detected.

name :: Maybe Text

The name of the custom data identifier.

newCustomDetection :: CustomDetection

Create a value of CustomDetection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:occurrences:CustomDetection', customDetection_occurrences - The location of 1-15 occurrences of the sensitive data that the custom data identifier detected. A finding includes location data for a maximum of 15 occurrences of sensitive data.

$sel:arn:CustomDetection', customDetection_arn - The Amazon Resource Name (ARN) of the custom data identifier.

$sel:count:CustomDetection', customDetection_count - The total number of occurrences of the sensitive data that the custom data identifier detected.

$sel:name:CustomDetection', customDetection_name - The name of the custom data identifier.

customDetection_occurrences :: Lens' CustomDetection (Maybe Occurrences)

The location of 1-15 occurrences of the sensitive data that the custom data identifier detected. A finding includes location data for a maximum of 15 occurrences of sensitive data.

customDetection_arn :: Lens' CustomDetection (Maybe Text)

The Amazon Resource Name (ARN) of the custom data identifier.

customDetection_count :: Lens' CustomDetection (Maybe Integer)

The total number of occurrences of the sensitive data that the custom data identifier detected.

customDetection_name :: Lens' CustomDetection (Maybe Text)

The name of the custom data identifier.