Network.AWS.MacieV2.Types.FederatedUser

Description

Documentation

data FederatedUser

Provides information about an identity that performed an action on an affected resource by using temporary security credentials. The credentials were obtained using the GetFederationToken operation of the Security Token Service (STS) API.

See: newFederatedUser smart constructor.

Constructors

FederatedUser'

Fields

principalId :: Maybe Text

The unique identifier for the entity that was used to get the credentials.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the entity that was used to get the credentials.

sessionContext :: Maybe SessionContext

The details of the session that was created for the credentials, including the entity that issued the session.

accountId :: Maybe Text

The unique identifier for the Amazon Web Services account that owns the entity that was used to get the credentials.

accessKeyId :: Maybe Text

The Amazon Web Services access key ID that identifies the credentials.

newFederatedUser :: FederatedUser

Create a value of FederatedUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:principalId:FederatedUser', federatedUser_principalId - The unique identifier for the entity that was used to get the credentials.

$sel:arn:FederatedUser', federatedUser_arn - The Amazon Resource Name (ARN) of the entity that was used to get the credentials.

$sel:sessionContext:FederatedUser', federatedUser_sessionContext - The details of the session that was created for the credentials, including the entity that issued the session.

$sel:accountId:FederatedUser', federatedUser_accountId - The unique identifier for the Amazon Web Services account that owns the entity that was used to get the credentials.

$sel:accessKeyId:FederatedUser', federatedUser_accessKeyId - The Amazon Web Services access key ID that identifies the credentials.

federatedUser_principalId :: Lens' FederatedUser (Maybe Text)

The unique identifier for the entity that was used to get the credentials.

federatedUser_arn :: Lens' FederatedUser (Maybe Text)

The Amazon Resource Name (ARN) of the entity that was used to get the credentials.

federatedUser_sessionContext :: Lens' FederatedUser (Maybe SessionContext)

The details of the session that was created for the credentials, including the entity that issued the session.

federatedUser_accountId :: Lens' FederatedUser (Maybe Text)

The unique identifier for the Amazon Web Services account that owns the entity that was used to get the credentials.

federatedUser_accessKeyId :: Lens' FederatedUser (Maybe Text)

The Amazon Web Services access key ID that identifies the credentials.