Network.AWS.MacieV2.Types.Finding

Description

Documentation

data Finding

Provides the details of a finding.

See: newFinding smart constructor.

Constructors

Finding'

Fields

classificationDetails :: Maybe ClassificationDetails

The details of a sensitive data finding. This value is null for a policy finding.

policyDetails :: Maybe PolicyDetails

The details of a policy finding. This value is null for a sensitive data finding.

createdAt :: Maybe POSIX

The date and time, in UTC and extended ISO 8601 format, when the finding was created.

category :: Maybe FindingCategory

The category of the finding. Possible values are: CLASSIFICATION, for a sensitive data finding; and, POLICY, for a policy finding.

severity :: Maybe Severity

The severity level and score for the finding.

count :: Maybe Integer

The total number of occurrences of the finding. For sensitive data findings, this value is always 1. All sensitive data findings are considered new (unique) because they derive from individual classification jobs.

schemaVersion :: Maybe Text

The version of the schema that was used to define the data structures in the finding.

resourcesAffected :: Maybe ResourcesAffected

The resources that the finding applies to.

accountId :: Maybe Text

The unique identifier for the Amazon Web Services account that the finding applies to. This is typically the account that owns the affected resource.

partition :: Maybe Text

The Amazon Web Services partition that Amazon Macie created the finding in.

id :: Maybe Text

The unique identifier for the finding. This is a random string that Amazon Macie generates and assigns to a finding when it creates the finding.

region :: Maybe Text

The Amazon Web Services Region that Amazon Macie created the finding in.

updatedAt :: Maybe POSIX

The date and time, in UTC and extended ISO 8601 format, when the finding was last updated. For sensitive data findings, this value is the same as the value for the createdAt property. All sensitive data findings are considered new (unique) because they derive from individual classification jobs.

title :: Maybe Text

The brief description of the finding.

type' :: Maybe FindingType

The type of the finding.

archived :: Maybe Bool

Specifies whether the finding is archived (suppressed).

description :: Maybe Text

The description of the finding.

sample :: Maybe Bool

Specifies whether the finding is a sample finding. A sample finding is a finding that uses example data to demonstrate what a finding might contain.

newFinding :: Finding

Create a value of Finding with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:classificationDetails:Finding', finding_classificationDetails - The details of a sensitive data finding. This value is null for a policy finding.

$sel:policyDetails:Finding', finding_policyDetails - The details of a policy finding. This value is null for a sensitive data finding.

$sel:createdAt:Finding', finding_createdAt - The date and time, in UTC and extended ISO 8601 format, when the finding was created.

$sel:category:Finding', finding_category - The category of the finding. Possible values are: CLASSIFICATION, for a sensitive data finding; and, POLICY, for a policy finding.

$sel:severity:Finding', finding_severity - The severity level and score for the finding.

$sel:count:Finding', finding_count - The total number of occurrences of the finding. For sensitive data findings, this value is always 1. All sensitive data findings are considered new (unique) because they derive from individual classification jobs.

$sel:schemaVersion:Finding', finding_schemaVersion - The version of the schema that was used to define the data structures in the finding.

$sel:resourcesAffected:Finding', finding_resourcesAffected - The resources that the finding applies to.

$sel:accountId:Finding', finding_accountId - The unique identifier for the Amazon Web Services account that the finding applies to. This is typically the account that owns the affected resource.

$sel:partition:Finding', finding_partition - The Amazon Web Services partition that Amazon Macie created the finding in.

$sel:id:Finding', finding_id - The unique identifier for the finding. This is a random string that Amazon Macie generates and assigns to a finding when it creates the finding.

$sel:region:Finding', finding_region - The Amazon Web Services Region that Amazon Macie created the finding in.

$sel:updatedAt:Finding', finding_updatedAt - The date and time, in UTC and extended ISO 8601 format, when the finding was last updated. For sensitive data findings, this value is the same as the value for the createdAt property. All sensitive data findings are considered new (unique) because they derive from individual classification jobs.

$sel:title:Finding', finding_title - The brief description of the finding.

$sel:type':Finding', finding_type - The type of the finding.

$sel:archived:Finding', finding_archived - Specifies whether the finding is archived (suppressed).

$sel:description:Finding', finding_description - The description of the finding.

$sel:sample:Finding', finding_sample - Specifies whether the finding is a sample finding. A sample finding is a finding that uses example data to demonstrate what a finding might contain.

finding_classificationDetails :: Lens' Finding (Maybe ClassificationDetails)

The details of a sensitive data finding. This value is null for a policy finding.

finding_policyDetails :: Lens' Finding (Maybe PolicyDetails)

The details of a policy finding. This value is null for a sensitive data finding.

finding_createdAt :: Lens' Finding (Maybe UTCTime)

The date and time, in UTC and extended ISO 8601 format, when the finding was created.

finding_category :: Lens' Finding (Maybe FindingCategory)

The category of the finding. Possible values are: CLASSIFICATION, for a sensitive data finding; and, POLICY, for a policy finding.

finding_severity :: Lens' Finding (Maybe Severity)

The severity level and score for the finding.

finding_count :: Lens' Finding (Maybe Integer)

The total number of occurrences of the finding. For sensitive data findings, this value is always 1. All sensitive data findings are considered new (unique) because they derive from individual classification jobs.

finding_schemaVersion :: Lens' Finding (Maybe Text)

The version of the schema that was used to define the data structures in the finding.

finding_resourcesAffected :: Lens' Finding (Maybe ResourcesAffected)

The resources that the finding applies to.

finding_accountId :: Lens' Finding (Maybe Text)

The unique identifier for the Amazon Web Services account that the finding applies to. This is typically the account that owns the affected resource.

finding_partition :: Lens' Finding (Maybe Text)

The Amazon Web Services partition that Amazon Macie created the finding in.

finding_id :: Lens' Finding (Maybe Text)

The unique identifier for the finding. This is a random string that Amazon Macie generates and assigns to a finding when it creates the finding.

finding_region :: Lens' Finding (Maybe Text)

The Amazon Web Services Region that Amazon Macie created the finding in.

finding_updatedAt :: Lens' Finding (Maybe UTCTime)

The date and time, in UTC and extended ISO 8601 format, when the finding was last updated. For sensitive data findings, this value is the same as the value for the createdAt property. All sensitive data findings are considered new (unique) because they derive from individual classification jobs.

finding_title :: Lens' Finding (Maybe Text)

The brief description of the finding.

finding_type :: Lens' Finding (Maybe FindingType)

The type of the finding.

finding_archived :: Lens' Finding (Maybe Bool)

Specifies whether the finding is archived (suppressed).

finding_description :: Lens' Finding (Maybe Text)

The description of the finding.

finding_sample :: Lens' Finding (Maybe Bool)

Specifies whether the finding is a sample finding. A sample finding is a finding that uses example data to demonstrate what a finding might contain.