Network.AWS.MacieV2.Types.FindingAction

Description

Documentation

data FindingAction

Provides information about an action that occurred for a resource and produced a policy finding.

See: newFindingAction smart constructor.

Constructors

FindingAction'

Fields

apiCallDetails :: Maybe ApiCallDetails

The invocation details of the API operation that an entity invoked for the affected resource, if the value for the actionType property is AWS_API_CALL.

actionType :: Maybe FindingActionType

The type of action that occurred for the affected resource. This value is typically AWS_API_CALL, which indicates that an entity invoked an API operation for the resource.

newFindingAction :: FindingAction

Create a value of FindingAction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apiCallDetails:FindingAction', findingAction_apiCallDetails - The invocation details of the API operation that an entity invoked for the affected resource, if the value for the actionType property is AWS_API_CALL.

$sel:actionType:FindingAction', findingAction_actionType - The type of action that occurred for the affected resource. This value is typically AWS_API_CALL, which indicates that an entity invoked an API operation for the resource.

findingAction_apiCallDetails :: Lens' FindingAction (Maybe ApiCallDetails)

The invocation details of the API operation that an entity invoked for the affected resource, if the value for the actionType property is AWS_API_CALL.

findingAction_actionType :: Lens' FindingAction (Maybe FindingActionType)

The type of action that occurred for the affected resource. This value is typically AWS_API_CALL, which indicates that an entity invoked an API operation for the resource.