Network.AWS.MacieV2.Types.IpAddressDetails

Description

Documentation

data IpAddressDetails

Provides information about the IP address of the device that an entity used to perform an action on an affected resource.

See: newIpAddressDetails smart constructor.

Constructors

IpAddressDetails'

Fields

ipCity :: Maybe IpCity

The city that the IP address originated from.

ipGeoLocation :: Maybe IpGeoLocation

The geographic coordinates of the location that the IP address originated from.

ipAddressV4 :: Maybe Text

The Internet Protocol version 4 (IPv4) address of the device.

ipOwner :: Maybe IpOwner

The registered owner of the IP address.

ipCountry :: Maybe IpCountry

The country that the IP address originated from.

newIpAddressDetails :: IpAddressDetails

Create a value of IpAddressDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ipCity:IpAddressDetails', ipAddressDetails_ipCity - The city that the IP address originated from.

$sel:ipGeoLocation:IpAddressDetails', ipAddressDetails_ipGeoLocation - The geographic coordinates of the location that the IP address originated from.

$sel:ipAddressV4:IpAddressDetails', ipAddressDetails_ipAddressV4 - The Internet Protocol version 4 (IPv4) address of the device.

$sel:ipOwner:IpAddressDetails', ipAddressDetails_ipOwner - The registered owner of the IP address.

$sel:ipCountry:IpAddressDetails', ipAddressDetails_ipCountry - The country that the IP address originated from.

ipAddressDetails_ipCity :: Lens' IpAddressDetails (Maybe IpCity)

The city that the IP address originated from.

ipAddressDetails_ipGeoLocation :: Lens' IpAddressDetails (Maybe IpGeoLocation)

The geographic coordinates of the location that the IP address originated from.

ipAddressDetails_ipAddressV4 :: Lens' IpAddressDetails (Maybe Text)

The Internet Protocol version 4 (IPv4) address of the device.

ipAddressDetails_ipOwner :: Lens' IpAddressDetails (Maybe IpOwner)

The registered owner of the IP address.

ipAddressDetails_ipCountry :: Lens' IpAddressDetails (Maybe IpCountry)

The country that the IP address originated from.