Network.AWS.MacieV2.Types.JobSummary

Description

Documentation

data JobSummary

Provides information about a classification job, including the current status of the job.

See: newJobSummary smart constructor.

Constructors

JobSummary'

Fields

lastRunErrorStatus :: Maybe LastRunErrorStatus

Specifies whether any account- or bucket-level access errors occurred when the job ran. For a recurring job, this value indicates the error status of the job's most recent run.

jobType :: Maybe JobType

The schedule for running the job. Possible values are:

ONE_TIME - The job runs only once.

SCHEDULED - The job runs on a daily, weekly, or monthly basis.

jobId :: Maybe Text

The unique identifier for the job.

createdAt :: Maybe POSIX

The date and time, in UTC and extended ISO 8601 format, when the job was created.

userPausedDetails :: Maybe UserPausedDetails

If the current status of the job is USER_PAUSED, specifies when the job was paused and when the job or job run will expire and be cancelled if it isn't resumed. This value is present only if the value for jobStatus is USER_PAUSED.

bucketCriteria :: Maybe S3BucketCriteriaForJob

The property- and tag-based conditions that determine which S3 buckets are included or excluded from the job's analysis. Each time the job runs, the job uses these criteria to determine which buckets to analyze. A job's definition can contain a bucketCriteria object or a bucketDefinitions array, not both.

name :: Maybe Text

The custom name of the job.

bucketDefinitions :: Maybe [S3BucketDefinitionForJob]

An array of objects, one for each Amazon Web Services account that owns specific S3 buckets for the job to analyze. Each object specifies the account ID for an account and one or more buckets to analyze for that account. A job's definition can contain a bucketDefinitions array or a bucketCriteria object, not both.

jobStatus :: Maybe JobStatus

The current status of the job. Possible values are:

CANCELLED - You cancelled the job or, if it's a one-time job, you paused the job and didn't resume it within 30 days.

COMPLETE - For a one-time job, Amazon Macie finished processing the data specified for the job. This value doesn't apply to recurring jobs.

IDLE - For a recurring job, the previous scheduled run is complete and the next scheduled run is pending. This value doesn't apply to one-time jobs.

PAUSED - Macie started running the job but additional processing would exceed the monthly sensitive data discovery quota for your account or one or more member accounts that the job analyzes data for.

RUNNING - For a one-time job, the job is in progress. For a recurring job, a scheduled run is in progress.

USER_PAUSED - You paused the job. If you paused the job while it had a status of RUNNING and you don't resume it within 30 days of pausing it, the job or job run will expire and be cancelled, depending on the job's type. To check the expiration date, refer to the UserPausedDetails.jobExpiresAt property.

newJobSummary :: JobSummary

Create a value of JobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastRunErrorStatus:JobSummary', jobSummary_lastRunErrorStatus - Specifies whether any account- or bucket-level access errors occurred when the job ran. For a recurring job, this value indicates the error status of the job's most recent run.

$sel:jobType:JobSummary', jobSummary_jobType - The schedule for running the job. Possible values are:

ONE_TIME - The job runs only once.

SCHEDULED - The job runs on a daily, weekly, or monthly basis.

$sel:jobId:JobSummary', jobSummary_jobId - The unique identifier for the job.

$sel:createdAt:JobSummary', jobSummary_createdAt - The date and time, in UTC and extended ISO 8601 format, when the job was created.

$sel:userPausedDetails:JobSummary', jobSummary_userPausedDetails - If the current status of the job is USER_PAUSED, specifies when the job was paused and when the job or job run will expire and be cancelled if it isn't resumed. This value is present only if the value for jobStatus is USER_PAUSED.

$sel:bucketCriteria:JobSummary', jobSummary_bucketCriteria - The property- and tag-based conditions that determine which S3 buckets are included or excluded from the job's analysis. Each time the job runs, the job uses these criteria to determine which buckets to analyze. A job's definition can contain a bucketCriteria object or a bucketDefinitions array, not both.

$sel:name:JobSummary', jobSummary_name - The custom name of the job.

$sel:bucketDefinitions:JobSummary', jobSummary_bucketDefinitions - An array of objects, one for each Amazon Web Services account that owns specific S3 buckets for the job to analyze. Each object specifies the account ID for an account and one or more buckets to analyze for that account. A job's definition can contain a bucketDefinitions array or a bucketCriteria object, not both.

$sel:jobStatus:JobSummary', jobSummary_jobStatus - The current status of the job. Possible values are:

CANCELLED - You cancelled the job or, if it's a one-time job, you paused the job and didn't resume it within 30 days.

COMPLETE - For a one-time job, Amazon Macie finished processing the data specified for the job. This value doesn't apply to recurring jobs.

IDLE - For a recurring job, the previous scheduled run is complete and the next scheduled run is pending. This value doesn't apply to one-time jobs.

PAUSED - Macie started running the job but additional processing would exceed the monthly sensitive data discovery quota for your account or one or more member accounts that the job analyzes data for.

RUNNING - For a one-time job, the job is in progress. For a recurring job, a scheduled run is in progress.

USER_PAUSED - You paused the job. If you paused the job while it had a status of RUNNING and you don't resume it within 30 days of pausing it, the job or job run will expire and be cancelled, depending on the job's type. To check the expiration date, refer to the UserPausedDetails.jobExpiresAt property.

jobSummary_lastRunErrorStatus :: Lens' JobSummary (Maybe LastRunErrorStatus)

Specifies whether any account- or bucket-level access errors occurred when the job ran. For a recurring job, this value indicates the error status of the job's most recent run.

jobSummary_jobType :: Lens' JobSummary (Maybe JobType)

The schedule for running the job. Possible values are:

ONE_TIME - The job runs only once.

SCHEDULED - The job runs on a daily, weekly, or monthly basis.

jobSummary_jobId :: Lens' JobSummary (Maybe Text)

The unique identifier for the job.

jobSummary_createdAt :: Lens' JobSummary (Maybe UTCTime)

The date and time, in UTC and extended ISO 8601 format, when the job was created.

jobSummary_userPausedDetails :: Lens' JobSummary (Maybe UserPausedDetails)

If the current status of the job is USER_PAUSED, specifies when the job was paused and when the job or job run will expire and be cancelled if it isn't resumed. This value is present only if the value for jobStatus is USER_PAUSED.

jobSummary_bucketCriteria :: Lens' JobSummary (Maybe S3BucketCriteriaForJob)

The property- and tag-based conditions that determine which S3 buckets are included or excluded from the job's analysis. Each time the job runs, the job uses these criteria to determine which buckets to analyze. A job's definition can contain a bucketCriteria object or a bucketDefinitions array, not both.

jobSummary_name :: Lens' JobSummary (Maybe Text)

The custom name of the job.

jobSummary_bucketDefinitions :: Lens' JobSummary (Maybe [S3BucketDefinitionForJob])

An array of objects, one for each Amazon Web Services account that owns specific S3 buckets for the job to analyze. Each object specifies the account ID for an account and one or more buckets to analyze for that account. A job's definition can contain a bucketDefinitions array or a bucketCriteria object, not both.

jobSummary_jobStatus :: Lens' JobSummary (Maybe JobStatus)

The current status of the job. Possible values are:

CANCELLED - You cancelled the job or, if it's a one-time job, you paused the job and didn't resume it within 30 days.

COMPLETE - For a one-time job, Amazon Macie finished processing the data specified for the job. This value doesn't apply to recurring jobs.

IDLE - For a recurring job, the previous scheduled run is complete and the next scheduled run is pending. This value doesn't apply to one-time jobs.

PAUSED - Macie started running the job but additional processing would exceed the monthly sensitive data discovery quota for your account or one or more member accounts that the job analyzes data for.

RUNNING - For a one-time job, the job is in progress. For a recurring job, a scheduled run is in progress.

USER_PAUSED - You paused the job. If you paused the job while it had a status of RUNNING and you don't resume it within 30 days of pausing it, the job or job run will expire and be cancelled, depending on the job's type. To check the expiration date, refer to the UserPausedDetails.jobExpiresAt property.