Network.AWS.MacieV2.Types.MatchingBucket

Description

Documentation

data MatchingBucket

Provides statistical data and other information about an S3 bucket that Amazon Macie monitors and analyzes for your account. If an error occurs when Macie attempts to retrieve and process information about the bucket or the bucket's objects, the value for most of these properties is null. Exceptions are accountId and bucketName. To identify the cause of the error, refer to the errorCode and errorMessage values.

See: newMatchingBucket smart constructor.

Constructors

MatchingBucket'

Fields

sizeInBytesCompressed :: Maybe Integer

The total storage size, in bytes, of the objects that are compressed (.gz, .gzip, .zip) files in the bucket.

If versioning is enabled for the bucket, Amazon Macie calculates this value based on the size of the latest version of each applicable object in the bucket. This value doesn't reflect the storage size of all versions of each applicable object in the bucket.

objectCountByEncryptionType :: Maybe ObjectCountByEncryptionType

The total number of objects that are in the bucket, grouped by server-side encryption type. This includes a grouping that reports the total number of objects that aren't encrypted or use client-side encryption.

sizeInBytes :: Maybe Integer

The total storage size, in bytes, of the bucket.

If versioning is enabled for the bucket, Amazon Macie calculates this value based on the size of the latest version of each object in the bucket. This value doesn't reflect the storage size of all versions of each object in the bucket.

bucketName :: Maybe Text

The name of the bucket.

classifiableObjectCount :: Maybe Integer

The total number of objects that Amazon Macie can analyze in the bucket. These objects use a supported storage class and have a file name extension for a supported file or storage format.

accountId :: Maybe Text

The unique identifier for the Amazon Web Services account that owns the bucket.

unclassifiableObjectSizeInBytes :: Maybe ObjectLevelStatistics

The total storage size, in bytes, of the objects that Amazon Macie can't analyze in the bucket. These objects don't use a supported storage class or don't have a file name extension for a supported file or storage format.

unclassifiableObjectCount :: Maybe ObjectLevelStatistics

The total number of objects that Amazon Macie can't analyze in the bucket. These objects don't use a supported storage class or don't have a file name extension for a supported file or storage format.

errorCode :: Maybe BucketMetadataErrorCode

Specifies the error code for an error that prevented Amazon Macie from retrieving and processing information about the bucket and the bucket's objects. If this value is ACCESS_DENIED, Macie doesn't have permission to retrieve the information. For example, the bucket has a restrictive bucket policy and Amazon S3 denied the request. If this value is null, Macie was able to retrieve and process the information.

jobDetails :: Maybe JobDetails

Specifies whether any one-time or recurring classification jobs are configured to analyze objects in the bucket, and, if so, the details of the job that ran most recently.

errorMessage :: Maybe Text

A brief description of the error (errorCode) that prevented Amazon Macie from retrieving and processing information about the bucket and the bucket's objects. This value is null if Macie was able to retrieve and process the information.

objectCount :: Maybe Integer

The total number of objects in the bucket.

classifiableSizeInBytes :: Maybe Integer

The total storage size, in bytes, of the objects that Amazon Macie can analyze in the bucket. These objects use a supported storage class and have a file name extension for a supported file or storage format.

If versioning is enabled for the bucket, Macie calculates this value based on the size of the latest version of each applicable object in the bucket. This value doesn't reflect the storage size of all versions of each applicable object in the bucket.

newMatchingBucket :: MatchingBucket

Create a value of MatchingBucket with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sizeInBytesCompressed:MatchingBucket', matchingBucket_sizeInBytesCompressed - The total storage size, in bytes, of the objects that are compressed (.gz, .gzip, .zip) files in the bucket.

If versioning is enabled for the bucket, Amazon Macie calculates this value based on the size of the latest version of each applicable object in the bucket. This value doesn't reflect the storage size of all versions of each applicable object in the bucket.

$sel:objectCountByEncryptionType:MatchingBucket', matchingBucket_objectCountByEncryptionType - The total number of objects that are in the bucket, grouped by server-side encryption type. This includes a grouping that reports the total number of objects that aren't encrypted or use client-side encryption.

$sel:sizeInBytes:MatchingBucket', matchingBucket_sizeInBytes - The total storage size, in bytes, of the bucket.

If versioning is enabled for the bucket, Amazon Macie calculates this value based on the size of the latest version of each object in the bucket. This value doesn't reflect the storage size of all versions of each object in the bucket.

$sel:bucketName:MatchingBucket', matchingBucket_bucketName - The name of the bucket.

$sel:classifiableObjectCount:MatchingBucket', matchingBucket_classifiableObjectCount - The total number of objects that Amazon Macie can analyze in the bucket. These objects use a supported storage class and have a file name extension for a supported file or storage format.

$sel:accountId:MatchingBucket', matchingBucket_accountId - The unique identifier for the Amazon Web Services account that owns the bucket.

$sel:unclassifiableObjectSizeInBytes:MatchingBucket', matchingBucket_unclassifiableObjectSizeInBytes - The total storage size, in bytes, of the objects that Amazon Macie can't analyze in the bucket. These objects don't use a supported storage class or don't have a file name extension for a supported file or storage format.

$sel:unclassifiableObjectCount:MatchingBucket', matchingBucket_unclassifiableObjectCount - The total number of objects that Amazon Macie can't analyze in the bucket. These objects don't use a supported storage class or don't have a file name extension for a supported file or storage format.

$sel:errorCode:MatchingBucket', matchingBucket_errorCode - Specifies the error code for an error that prevented Amazon Macie from retrieving and processing information about the bucket and the bucket's objects. If this value is ACCESS_DENIED, Macie doesn't have permission to retrieve the information. For example, the bucket has a restrictive bucket policy and Amazon S3 denied the request. If this value is null, Macie was able to retrieve and process the information.

$sel:jobDetails:MatchingBucket', matchingBucket_jobDetails - Specifies whether any one-time or recurring classification jobs are configured to analyze objects in the bucket, and, if so, the details of the job that ran most recently.

$sel:errorMessage:MatchingBucket', matchingBucket_errorMessage - A brief description of the error (errorCode) that prevented Amazon Macie from retrieving and processing information about the bucket and the bucket's objects. This value is null if Macie was able to retrieve and process the information.

$sel:objectCount:MatchingBucket', matchingBucket_objectCount - The total number of objects in the bucket.

$sel:classifiableSizeInBytes:MatchingBucket', matchingBucket_classifiableSizeInBytes - The total storage size, in bytes, of the objects that Amazon Macie can analyze in the bucket. These objects use a supported storage class and have a file name extension for a supported file or storage format.

If versioning is enabled for the bucket, Macie calculates this value based on the size of the latest version of each applicable object in the bucket. This value doesn't reflect the storage size of all versions of each applicable object in the bucket.

matchingBucket_sizeInBytesCompressed :: Lens' MatchingBucket (Maybe Integer)

The total storage size, in bytes, of the objects that are compressed (.gz, .gzip, .zip) files in the bucket.

If versioning is enabled for the bucket, Amazon Macie calculates this value based on the size of the latest version of each applicable object in the bucket. This value doesn't reflect the storage size of all versions of each applicable object in the bucket.

matchingBucket_objectCountByEncryptionType :: Lens' MatchingBucket (Maybe ObjectCountByEncryptionType)

The total number of objects that are in the bucket, grouped by server-side encryption type. This includes a grouping that reports the total number of objects that aren't encrypted or use client-side encryption.

matchingBucket_sizeInBytes :: Lens' MatchingBucket (Maybe Integer)

The total storage size, in bytes, of the bucket.

If versioning is enabled for the bucket, Amazon Macie calculates this value based on the size of the latest version of each object in the bucket. This value doesn't reflect the storage size of all versions of each object in the bucket.

matchingBucket_bucketName :: Lens' MatchingBucket (Maybe Text)

The name of the bucket.

matchingBucket_classifiableObjectCount :: Lens' MatchingBucket (Maybe Integer)

The total number of objects that Amazon Macie can analyze in the bucket. These objects use a supported storage class and have a file name extension for a supported file or storage format.

matchingBucket_accountId :: Lens' MatchingBucket (Maybe Text)

The unique identifier for the Amazon Web Services account that owns the bucket.

matchingBucket_unclassifiableObjectSizeInBytes :: Lens' MatchingBucket (Maybe ObjectLevelStatistics)

The total storage size, in bytes, of the objects that Amazon Macie can't analyze in the bucket. These objects don't use a supported storage class or don't have a file name extension for a supported file or storage format.

matchingBucket_unclassifiableObjectCount :: Lens' MatchingBucket (Maybe ObjectLevelStatistics)

The total number of objects that Amazon Macie can't analyze in the bucket. These objects don't use a supported storage class or don't have a file name extension for a supported file or storage format.

matchingBucket_errorCode :: Lens' MatchingBucket (Maybe BucketMetadataErrorCode)

Specifies the error code for an error that prevented Amazon Macie from retrieving and processing information about the bucket and the bucket's objects. If this value is ACCESS_DENIED, Macie doesn't have permission to retrieve the information. For example, the bucket has a restrictive bucket policy and Amazon S3 denied the request. If this value is null, Macie was able to retrieve and process the information.

matchingBucket_jobDetails :: Lens' MatchingBucket (Maybe JobDetails)

Specifies whether any one-time or recurring classification jobs are configured to analyze objects in the bucket, and, if so, the details of the job that ran most recently.

matchingBucket_errorMessage :: Lens' MatchingBucket (Maybe Text)

A brief description of the error (errorCode) that prevented Amazon Macie from retrieving and processing information about the bucket and the bucket's objects. This value is null if Macie was able to retrieve and process the information.

matchingBucket_objectCount :: Lens' MatchingBucket (Maybe Integer)

The total number of objects in the bucket.

matchingBucket_classifiableSizeInBytes :: Lens' MatchingBucket (Maybe Integer)

The total storage size, in bytes, of the objects that Amazon Macie can analyze in the bucket. These objects use a supported storage class and have a file name extension for a supported file or storage format.

If versioning is enabled for the bucket, Macie calculates this value based on the size of the latest version of each applicable object in the bucket. This value doesn't reflect the storage size of all versions of each applicable object in the bucket.