Network.AWS.MacieV2.Types.Member

Description

Documentation

data Member

Provides information about an account that's associated with an Amazon Macie administrator account.

See: newMember smart constructor.

Constructors

Member'

Fields

email :: Maybe Text

The email address for the account.

administratorAccountId :: Maybe Text

The Amazon Web Services account ID for the administrator account.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the account.

relationshipStatus :: Maybe RelationshipStatus

The current status of the relationship between the account and the administrator account.

masterAccountId :: Maybe Text

(Deprecated) The Amazon Web Services account ID for the administrator account. This property has been replaced by the administratorAccountId property and is retained only for backward compatibility.

invitedAt :: Maybe POSIX

The date and time, in UTC and extended ISO 8601 format, when an Amazon Macie membership invitation was last sent to the account. This value is null if a Macie invitation hasn't been sent to the account.

accountId :: Maybe Text

The Amazon Web Services account ID for the account.

updatedAt :: Maybe POSIX

The date and time, in UTC and extended ISO 8601 format, of the most recent change to the status of the relationship between the account and the administrator account.

tags :: Maybe (HashMap Text Text)

A map of key-value pairs that identifies the tags (keys and values) that are associated with the account in Amazon Macie.

newMember :: Member

Create a value of Member with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:email:Member', member_email - The email address for the account.

$sel:administratorAccountId:Member', member_administratorAccountId - The Amazon Web Services account ID for the administrator account.

$sel:arn:Member', member_arn - The Amazon Resource Name (ARN) of the account.

$sel:relationshipStatus:Member', member_relationshipStatus - The current status of the relationship between the account and the administrator account.

$sel:masterAccountId:Member', member_masterAccountId - (Deprecated) The Amazon Web Services account ID for the administrator account. This property has been replaced by the administratorAccountId property and is retained only for backward compatibility.

$sel:invitedAt:Member', member_invitedAt - The date and time, in UTC and extended ISO 8601 format, when an Amazon Macie membership invitation was last sent to the account. This value is null if a Macie invitation hasn't been sent to the account.

$sel:accountId:Member', member_accountId - The Amazon Web Services account ID for the account.

$sel:updatedAt:Member', member_updatedAt - The date and time, in UTC and extended ISO 8601 format, of the most recent change to the status of the relationship between the account and the administrator account.

$sel:tags:Member', member_tags - A map of key-value pairs that identifies the tags (keys and values) that are associated with the account in Amazon Macie.

member_email :: Lens' Member (Maybe Text)

The email address for the account.

member_administratorAccountId :: Lens' Member (Maybe Text)

The Amazon Web Services account ID for the administrator account.

member_arn :: Lens' Member (Maybe Text)

The Amazon Resource Name (ARN) of the account.

member_relationshipStatus :: Lens' Member (Maybe RelationshipStatus)

The current status of the relationship between the account and the administrator account.

member_masterAccountId :: Lens' Member (Maybe Text)

(Deprecated) The Amazon Web Services account ID for the administrator account. This property has been replaced by the administratorAccountId property and is retained only for backward compatibility.

member_invitedAt :: Lens' Member (Maybe UTCTime)

The date and time, in UTC and extended ISO 8601 format, when an Amazon Macie membership invitation was last sent to the account. This value is null if a Macie invitation hasn't been sent to the account.

member_accountId :: Lens' Member (Maybe Text)

The Amazon Web Services account ID for the account.

member_updatedAt :: Lens' Member (Maybe UTCTime)

The date and time, in UTC and extended ISO 8601 format, of the most recent change to the status of the relationship between the account and the administrator account.

member_tags :: Lens' Member (Maybe (HashMap Text Text))

A map of key-value pairs that identifies the tags (keys and values) that are associated with the account in Amazon Macie.