Network.AWS.MacieV2.Types.S3Destination

Description

Documentation

data S3Destination

Specifies an S3 bucket to store data classification results in, and the encryption settings to use when storing results in that bucket.

See: newS3Destination smart constructor.

Constructors

S3Destination'

Fields

keyPrefix :: Maybe Text

The path prefix to use in the path to the location in the bucket. This prefix specifies where to store classification results in the bucket.

bucketName :: Text

The name of the bucket.

kmsKeyArn :: Text

The Amazon Resource Name (ARN) of the KMS key to use for encryption of the results. This must be the ARN of an existing, symmetric, customer managed KMS key that's in the same Amazon Web Services Region as the bucket.

newS3Destination

Arguments

:: Text
$sel:bucketName:S3Destination'

-> Text
$sel:kmsKeyArn:S3Destination'

-> S3Destination

Create a value of S3Destination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyPrefix:S3Destination', s3Destination_keyPrefix - The path prefix to use in the path to the location in the bucket. This prefix specifies where to store classification results in the bucket.

$sel:bucketName:S3Destination', s3Destination_bucketName - The name of the bucket.

$sel:kmsKeyArn:S3Destination', s3Destination_kmsKeyArn - The Amazon Resource Name (ARN) of the KMS key to use for encryption of the results. This must be the ARN of an existing, symmetric, customer managed KMS key that's in the same Amazon Web Services Region as the bucket.

s3Destination_keyPrefix :: Lens' S3Destination (Maybe Text)

The path prefix to use in the path to the location in the bucket. This prefix specifies where to store classification results in the bucket.

s3Destination_bucketName :: Lens' S3Destination Text

The name of the bucket.

s3Destination_kmsKeyArn :: Lens' S3Destination Text

The Amazon Resource Name (ARN) of the KMS key to use for encryption of the results. This must be the ARN of an existing, symmetric, customer managed KMS key that's in the same Amazon Web Services Region as the bucket.