Network.AWS.MacieV2.Types.SearchResourcesSortCriteria

Description

Documentation

data SearchResourcesSortCriteria

Specifies criteria for sorting the results of a query for information about Amazon Web Services resources that Amazon Macie monitors and analyzes.

See: newSearchResourcesSortCriteria smart constructor.

Constructors

SearchResourcesSortCriteria'

Fields

orderBy :: Maybe OrderBy

The sort order to apply to the results, based on the value for the property specified by the attributeName property. Valid values are: ASC, sort the results in ascending order; and, DESC, sort the results in descending order.

attributeName :: Maybe SearchResourcesSortAttributeName

The property to sort the results by.

newSearchResourcesSortCriteria :: SearchResourcesSortCriteria

Create a value of SearchResourcesSortCriteria with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:orderBy:SearchResourcesSortCriteria', searchResourcesSortCriteria_orderBy - The sort order to apply to the results, based on the value for the property specified by the attributeName property. Valid values are: ASC, sort the results in ascending order; and, DESC, sort the results in descending order.

$sel:attributeName:SearchResourcesSortCriteria', searchResourcesSortCriteria_attributeName - The property to sort the results by.

searchResourcesSortCriteria_orderBy :: Lens' SearchResourcesSortCriteria (Maybe OrderBy)

The sort order to apply to the results, based on the value for the property specified by the attributeName property. Valid values are: ASC, sort the results in ascending order; and, DESC, sort the results in descending order.

searchResourcesSortCriteria_attributeName :: Lens' SearchResourcesSortCriteria (Maybe SearchResourcesSortAttributeName)

The property to sort the results by.