Network.AWS.MacieV2.Types.SecurityHubConfiguration

Description

Documentation

data SecurityHubConfiguration

Specifies configuration settings that determine which findings are published to Security Hub automatically. For information about how Macie publishes findings to Security Hub, see Amazon Macie integration with Security Hub in the Amazon Macie User Guide.

See: newSecurityHubConfiguration smart constructor.

Constructors

SecurityHubConfiguration'

Fields

publishPolicyFindings :: Bool

Specifies whether to publish policy findings to Security Hub. If you set this value to true, Amazon Macie automatically publishes all new and updated policy findings that weren't suppressed by a findings filter. The default value is true.

publishClassificationFindings :: Bool

Specifies whether to publish sensitive data findings to Security Hub. If you set this value to true, Amazon Macie automatically publishes all sensitive data findings that weren't suppressed by a findings filter. The default value is false.

newSecurityHubConfiguration

Arguments

:: Bool
$sel:publishPolicyFindings:SecurityHubConfiguration'

-> Bool
$sel:publishClassificationFindings:SecurityHubConfiguration'

-> SecurityHubConfiguration

Create a value of SecurityHubConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:publishPolicyFindings:SecurityHubConfiguration', securityHubConfiguration_publishPolicyFindings - Specifies whether to publish policy findings to Security Hub. If you set this value to true, Amazon Macie automatically publishes all new and updated policy findings that weren't suppressed by a findings filter. The default value is true.

$sel:publishClassificationFindings:SecurityHubConfiguration', securityHubConfiguration_publishClassificationFindings - Specifies whether to publish sensitive data findings to Security Hub. If you set this value to true, Amazon Macie automatically publishes all sensitive data findings that weren't suppressed by a findings filter. The default value is false.

securityHubConfiguration_publishPolicyFindings :: Lens' SecurityHubConfiguration Bool

Specifies whether to publish policy findings to Security Hub. If you set this value to true, Amazon Macie automatically publishes all new and updated policy findings that weren't suppressed by a findings filter. The default value is true.

securityHubConfiguration_publishClassificationFindings :: Lens' SecurityHubConfiguration Bool

Specifies whether to publish sensitive data findings to Security Hub. If you set this value to true, Amazon Macie automatically publishes all sensitive data findings that weren't suppressed by a findings filter. The default value is false.