Network.AWS.MacieV2.Types.SensitiveDataItem

Description

Documentation

data SensitiveDataItem

Provides information about the category, types, and occurrences of sensitive data that produced a sensitive data finding.

See: newSensitiveDataItem smart constructor.

Constructors

SensitiveDataItem'

Fields

detections :: Maybe [DefaultDetection]

An array of objects, one for each type of sensitive data that was detected. Each object reports the number of occurrences of a specific type of sensitive data that was detected, and the location of up to 15 of those occurrences.

category :: Maybe SensitiveDataItemCategory

The category of sensitive data that was detected. For example: CREDENTIALS, for credentials data such as private keys or Amazon Web Services secret keys; FINANCIAL_INFORMATION, for financial data such as credit card numbers; or, PERSONAL_INFORMATION, for personal health information, such as health insurance identification numbers, or personally identifiable information, such as passport numbers.

totalCount :: Maybe Integer

The total number of occurrences of the sensitive data that was detected.

newSensitiveDataItem :: SensitiveDataItem

Create a value of SensitiveDataItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:detections:SensitiveDataItem', sensitiveDataItem_detections - An array of objects, one for each type of sensitive data that was detected. Each object reports the number of occurrences of a specific type of sensitive data that was detected, and the location of up to 15 of those occurrences.

$sel:category:SensitiveDataItem', sensitiveDataItem_category - The category of sensitive data that was detected. For example: CREDENTIALS, for credentials data such as private keys or Amazon Web Services secret keys; FINANCIAL_INFORMATION, for financial data such as credit card numbers; or, PERSONAL_INFORMATION, for personal health information, such as health insurance identification numbers, or personally identifiable information, such as passport numbers.

$sel:totalCount:SensitiveDataItem', sensitiveDataItem_totalCount - The total number of occurrences of the sensitive data that was detected.

sensitiveDataItem_detections :: Lens' SensitiveDataItem (Maybe [DefaultDetection])

An array of objects, one for each type of sensitive data that was detected. Each object reports the number of occurrences of a specific type of sensitive data that was detected, and the location of up to 15 of those occurrences.

sensitiveDataItem_category :: Lens' SensitiveDataItem (Maybe SensitiveDataItemCategory)

The category of sensitive data that was detected. For example: CREDENTIALS, for credentials data such as private keys or Amazon Web Services secret keys; FINANCIAL_INFORMATION, for financial data such as credit card numbers; or, PERSONAL_INFORMATION, for personal health information, such as health insurance identification numbers, or personally identifiable information, such as passport numbers.

sensitiveDataItem_totalCount :: Lens' SensitiveDataItem (Maybe Integer)

The total number of occurrences of the sensitive data that was detected.