Network.AWS.MacieV2.Types.SimpleCriterionForJob

Description

Documentation

data SimpleCriterionForJob

Specifies a property-based condition that determines whether an S3 bucket is included or excluded from a classification job.

See: newSimpleCriterionForJob smart constructor.

Constructors

SimpleCriterionForJob'

Fields

values :: Maybe [Text]

An array that lists one or more values to use in the condition. If you specify multiple values, Amazon Macie uses OR logic to join the values. Valid values for each supported property (key) are:

ACCOUNT_ID - A string that represents the unique identifier for the Amazon Web Services account that owns the bucket.

S3_BUCKET_EFFECTIVE_PERMISSION - A string that represents an enumerated value that Macie defines for the BucketPublicAccess.effectivePermission property of a bucket.

S3_BUCKET_NAME - A string that represents the name of a bucket.

S3_BUCKET_SHARED_ACCESS - A string that represents an enumerated value that Macie defines for the BucketMetadata.sharedAccess property of a bucket.

Values are case sensitive. Also, Macie doesn't support use of partial values or wildcard characters in these values.

key :: Maybe SimpleCriterionKeyForJob

The property to use in the condition.

comparator :: Maybe JobComparator

The operator to use in the condition. Valid values are EQ (equals) and NE (not equals).

newSimpleCriterionForJob :: SimpleCriterionForJob

Create a value of SimpleCriterionForJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:SimpleCriterionForJob', simpleCriterionForJob_values - An array that lists one or more values to use in the condition. If you specify multiple values, Amazon Macie uses OR logic to join the values. Valid values for each supported property (key) are:

ACCOUNT_ID - A string that represents the unique identifier for the Amazon Web Services account that owns the bucket.

S3_BUCKET_EFFECTIVE_PERMISSION - A string that represents an enumerated value that Macie defines for the BucketPublicAccess.effectivePermission property of a bucket.

S3_BUCKET_NAME - A string that represents the name of a bucket.

S3_BUCKET_SHARED_ACCESS - A string that represents an enumerated value that Macie defines for the BucketMetadata.sharedAccess property of a bucket.

Values are case sensitive. Also, Macie doesn't support use of partial values or wildcard characters in these values.

$sel:key:SimpleCriterionForJob', simpleCriterionForJob_key - The property to use in the condition.

$sel:comparator:SimpleCriterionForJob', simpleCriterionForJob_comparator - The operator to use in the condition. Valid values are EQ (equals) and NE (not equals).

simpleCriterionForJob_values :: Lens' SimpleCriterionForJob (Maybe [Text])

An array that lists one or more values to use in the condition. If you specify multiple values, Amazon Macie uses OR logic to join the values. Valid values for each supported property (key) are:

ACCOUNT_ID - A string that represents the unique identifier for the Amazon Web Services account that owns the bucket.

S3_BUCKET_EFFECTIVE_PERMISSION - A string that represents an enumerated value that Macie defines for the BucketPublicAccess.effectivePermission property of a bucket.

S3_BUCKET_NAME - A string that represents the name of a bucket.

S3_BUCKET_SHARED_ACCESS - A string that represents an enumerated value that Macie defines for the BucketMetadata.sharedAccess property of a bucket.

Values are case sensitive. Also, Macie doesn't support use of partial values or wildcard characters in these values.

simpleCriterionForJob_key :: Lens' SimpleCriterionForJob (Maybe SimpleCriterionKeyForJob)

The property to use in the condition.

simpleCriterionForJob_comparator :: Lens' SimpleCriterionForJob (Maybe JobComparator)

The operator to use in the condition. Valid values are EQ (equals) and NE (not equals).