Network.AWS.MacieV2.Types.TagCriterionForJob

Description

Documentation

data TagCriterionForJob

Specifies a tag-based condition that determines whether an S3 bucket is included or excluded from a classification job.

See: newTagCriterionForJob smart constructor.

Constructors

TagCriterionForJob'

Fields

tagValues :: Maybe [TagCriterionPairForJob]

The tag keys, tag values, or tag key and value pairs to use in the condition.

comparator :: Maybe JobComparator

The operator to use in the condition. Valid values are EQ (equals) and NE (not equals).

newTagCriterionForJob :: TagCriterionForJob

Create a value of TagCriterionForJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagValues:TagCriterionForJob', tagCriterionForJob_tagValues - The tag keys, tag values, or tag key and value pairs to use in the condition.

$sel:comparator:TagCriterionForJob', tagCriterionForJob_comparator - The operator to use in the condition. Valid values are EQ (equals) and NE (not equals).

tagCriterionForJob_tagValues :: Lens' TagCriterionForJob (Maybe [TagCriterionPairForJob])

The tag keys, tag values, or tag key and value pairs to use in the condition.

tagCriterionForJob_comparator :: Lens' TagCriterionForJob (Maybe JobComparator)

The operator to use in the condition. Valid values are EQ (equals) and NE (not equals).