Network.AWS.MacieV2.Types.TagScopeTerm

Description

Documentation

data TagScopeTerm

Specifies a tag-based condition that determines whether an S3 object is included or excluded from a classification job.

See: newTagScopeTerm smart constructor.

Constructors

TagScopeTerm'

Fields

tagValues :: Maybe [TagValuePair]

The tag keys or tag key and value pairs to use in the condition. To specify only tag keys in a condition, specify the keys in this array and set the value for each associated tag value to an empty string.

key :: Maybe Text

The object property to use in the condition. The only valid value is TAG.

comparator :: Maybe JobComparator

The operator to use in the condition. Valid values are EQ (equals) or NE (not equals).

target :: Maybe TagTarget

The type of object to apply the condition to.

newTagScopeTerm :: TagScopeTerm

Create a value of TagScopeTerm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagValues:TagScopeTerm', tagScopeTerm_tagValues - The tag keys or tag key and value pairs to use in the condition. To specify only tag keys in a condition, specify the keys in this array and set the value for each associated tag value to an empty string.

$sel:key:TagScopeTerm', tagScopeTerm_key - The object property to use in the condition. The only valid value is TAG.

$sel:comparator:TagScopeTerm', tagScopeTerm_comparator - The operator to use in the condition. Valid values are EQ (equals) or NE (not equals).

$sel:target:TagScopeTerm', tagScopeTerm_target - The type of object to apply the condition to.

tagScopeTerm_tagValues :: Lens' TagScopeTerm (Maybe [TagValuePair])

The tag keys or tag key and value pairs to use in the condition. To specify only tag keys in a condition, specify the keys in this array and set the value for each associated tag value to an empty string.

tagScopeTerm_key :: Lens' TagScopeTerm (Maybe Text)

The object property to use in the condition. The only valid value is TAG.

tagScopeTerm_comparator :: Lens' TagScopeTerm (Maybe JobComparator)

The operator to use in the condition. Valid values are EQ (equals) or NE (not equals).

tagScopeTerm_target :: Lens' TagScopeTerm (Maybe TagTarget)

The type of object to apply the condition to.