Network.AWS.MacieV2.Types.UsageRecord

Description

Documentation

data UsageRecord

Provides quota and aggregated usage data for an Amazon Macie account.

See: newUsageRecord smart constructor.

Constructors

UsageRecord'

Fields

accountId :: Maybe Text

The unique identifier for the Amazon Web Services account that the data applies to.

freeTrialStartDate :: Maybe POSIX

The date and time, in UTC and extended ISO 8601 format, when the free trial started for the account.

usage :: Maybe [UsageByAccount]

An array of objects that contains usage data and quotas for the account. Each object contains the data for a specific usage metric and the corresponding quota.

newUsageRecord :: UsageRecord

Create a value of UsageRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:UsageRecord', usageRecord_accountId - The unique identifier for the Amazon Web Services account that the data applies to.

$sel:freeTrialStartDate:UsageRecord', usageRecord_freeTrialStartDate - The date and time, in UTC and extended ISO 8601 format, when the free trial started for the account.

$sel:usage:UsageRecord', usageRecord_usage - An array of objects that contains usage data and quotas for the account. Each object contains the data for a specific usage metric and the corresponding quota.

usageRecord_accountId :: Lens' UsageRecord (Maybe Text)

The unique identifier for the Amazon Web Services account that the data applies to.

usageRecord_freeTrialStartDate :: Lens' UsageRecord (Maybe UTCTime)

The date and time, in UTC and extended ISO 8601 format, when the free trial started for the account.

usageRecord_usage :: Lens' UsageRecord (Maybe [UsageByAccount])

An array of objects that contains usage data and quotas for the account. Each object contains the data for a specific usage metric and the corresponding quota.