Network.AWS.MacieV2.Types.UsageStatisticsFilter

Description

Documentation

data UsageStatisticsFilter

Specifies a condition for filtering the results of a query for quota and usage data for one or more Amazon Macie accounts.

See: newUsageStatisticsFilter smart constructor.

Constructors

UsageStatisticsFilter'

Fields

values :: Maybe [Text]

An array that lists values to use in the condition, based on the value for the field specified by the key property. If the value for the key property is accountId, this array can specify multiple values. Otherwise, this array can specify only one value.

Valid values for each supported field are:

accountId - The unique identifier for an Amazon Web Services account.

freeTrialStartDate - The date and time, in UTC and extended ISO 8601 format, when the free trial started for an account.

serviceLimit - A Boolean (true or false) value that indicates whether an account has reached its monthly quota.

total - A string that represents the current estimated cost for an account.

key :: Maybe UsageStatisticsFilterKey

The field to use in the condition.

comparator :: Maybe UsageStatisticsFilterComparator

The operator to use in the condition. If the value for the key property is accountId, this value must be CONTAINS. If the value for the key property is any other supported field, this value can be EQ, GT, GTE, LT, LTE, or NE.

newUsageStatisticsFilter :: UsageStatisticsFilter

Create a value of UsageStatisticsFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:values:UsageStatisticsFilter', usageStatisticsFilter_values - An array that lists values to use in the condition, based on the value for the field specified by the key property. If the value for the key property is accountId, this array can specify multiple values. Otherwise, this array can specify only one value.

Valid values for each supported field are:

accountId - The unique identifier for an Amazon Web Services account.

freeTrialStartDate - The date and time, in UTC and extended ISO 8601 format, when the free trial started for an account.

serviceLimit - A Boolean (true or false) value that indicates whether an account has reached its monthly quota.

total - A string that represents the current estimated cost for an account.

$sel:key:UsageStatisticsFilter', usageStatisticsFilter_key - The field to use in the condition.

$sel:comparator:UsageStatisticsFilter', usageStatisticsFilter_comparator - The operator to use in the condition. If the value for the key property is accountId, this value must be CONTAINS. If the value for the key property is any other supported field, this value can be EQ, GT, GTE, LT, LTE, or NE.

usageStatisticsFilter_values :: Lens' UsageStatisticsFilter (Maybe [Text])

An array that lists values to use in the condition, based on the value for the field specified by the key property. If the value for the key property is accountId, this array can specify multiple values. Otherwise, this array can specify only one value.

Valid values for each supported field are:

accountId - The unique identifier for an Amazon Web Services account.

freeTrialStartDate - The date and time, in UTC and extended ISO 8601 format, when the free trial started for an account.

serviceLimit - A Boolean (true or false) value that indicates whether an account has reached its monthly quota.

total - A string that represents the current estimated cost for an account.

usageStatisticsFilter_key :: Lens' UsageStatisticsFilter (Maybe UsageStatisticsFilterKey)

The field to use in the condition.

usageStatisticsFilter_comparator :: Lens' UsageStatisticsFilter (Maybe UsageStatisticsFilterComparator)

The operator to use in the condition. If the value for the key property is accountId, this value must be CONTAINS. If the value for the key property is any other supported field, this value can be EQ, GT, GTE, LT, LTE, or NE.