Network.AWS.MacieV2.UpdateClassificationJob

Description

Changes the status of a classification job.

Creating a Request

data UpdateClassificationJob

See: newUpdateClassificationJob smart constructor.

Constructors

UpdateClassificationJob'

Fields

jobId :: Text

The unique identifier for the classification job.

jobStatus :: JobStatus

The new status for the job. Valid values are:

CANCELLED - Stops the job permanently and cancels it. This value is valid only if the job's current status is IDLE, PAUSED, RUNNING, or USER_PAUSED.

If you specify this value and the job's current status is RUNNING, Amazon Macie immediately begins to stop all processing tasks for the job. You can't resume or restart a job after you cancel it.

RUNNING - Resumes the job. This value is valid only if the job's current status is USER_PAUSED.

If you paused the job while it was actively running and you specify this value less than 30 days after you paused the job, Macie immediately resumes processing from the point where you paused the job. Otherwise, Macie resumes the job according to the schedule and other settings for the job.

USER_PAUSED - Pauses the job temporarily. This value is valid only if the job's current status is IDLE, PAUSED, or RUNNING. If you specify this value and the job's current status is RUNNING, Macie immediately begins to pause all processing tasks for the job.

If you pause a one-time job and you don't resume it within 30 days, the job expires and Macie cancels the job. If you pause a recurring job when its status is RUNNING and you don't resume it within 30 days, the job run expires and Macie cancels the run. To check the expiration date, refer to the UserPausedDetails.jobExpiresAt property.

newUpdateClassificationJob

Arguments

:: Text
$sel:jobId:UpdateClassificationJob'

-> JobStatus
$sel:jobStatus:UpdateClassificationJob'

-> UpdateClassificationJob

Create a value of UpdateClassificationJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:UpdateClassificationJob', updateClassificationJob_jobId - The unique identifier for the classification job.

$sel:jobStatus:UpdateClassificationJob', updateClassificationJob_jobStatus - The new status for the job. Valid values are:

CANCELLED - Stops the job permanently and cancels it. This value is valid only if the job's current status is IDLE, PAUSED, RUNNING, or USER_PAUSED.

If you specify this value and the job's current status is RUNNING, Amazon Macie immediately begins to stop all processing tasks for the job. You can't resume or restart a job after you cancel it.

RUNNING - Resumes the job. This value is valid only if the job's current status is USER_PAUSED.

If you paused the job while it was actively running and you specify this value less than 30 days after you paused the job, Macie immediately resumes processing from the point where you paused the job. Otherwise, Macie resumes the job according to the schedule and other settings for the job.

USER_PAUSED - Pauses the job temporarily. This value is valid only if the job's current status is IDLE, PAUSED, or RUNNING. If you specify this value and the job's current status is RUNNING, Macie immediately begins to pause all processing tasks for the job.

If you pause a one-time job and you don't resume it within 30 days, the job expires and Macie cancels the job. If you pause a recurring job when its status is RUNNING and you don't resume it within 30 days, the job run expires and Macie cancels the run. To check the expiration date, refer to the UserPausedDetails.jobExpiresAt property.

Request Lenses

updateClassificationJob_jobId :: Lens' UpdateClassificationJob Text

The unique identifier for the classification job.

updateClassificationJob_jobStatus :: Lens' UpdateClassificationJob JobStatus

The new status for the job. Valid values are:

CANCELLED - Stops the job permanently and cancels it. This value is valid only if the job's current status is IDLE, PAUSED, RUNNING, or USER_PAUSED.

If you specify this value and the job's current status is RUNNING, Amazon Macie immediately begins to stop all processing tasks for the job. You can't resume or restart a job after you cancel it.

RUNNING - Resumes the job. This value is valid only if the job's current status is USER_PAUSED.

If you paused the job while it was actively running and you specify this value less than 30 days after you paused the job, Macie immediately resumes processing from the point where you paused the job. Otherwise, Macie resumes the job according to the schedule and other settings for the job.

USER_PAUSED - Pauses the job temporarily. This value is valid only if the job's current status is IDLE, PAUSED, or RUNNING. If you specify this value and the job's current status is RUNNING, Macie immediately begins to pause all processing tasks for the job.

If you pause a one-time job and you don't resume it within 30 days, the job expires and Macie cancels the job. If you pause a recurring job when its status is RUNNING and you don't resume it within 30 days, the job run expires and Macie cancels the run. To check the expiration date, refer to the UserPausedDetails.jobExpiresAt property.

Destructuring the Response

data UpdateClassificationJobResponse

See: newUpdateClassificationJobResponse smart constructor.

Constructors

UpdateClassificationJobResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateClassificationJobResponse

Arguments

:: Int
$sel:httpStatus:UpdateClassificationJobResponse'

-> UpdateClassificationJobResponse

Create a value of UpdateClassificationJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateClassificationJobResponse', updateClassificationJobResponse_httpStatus - The response's http status code.

Response Lenses

updateClassificationJobResponse_httpStatus :: Lens' UpdateClassificationJobResponse Int

The response's http status code.