Network.AWS.MacieV2.UpdateFindingsFilter

Description

Updates the criteria and other settings for a findings filter.

Creating a Request

data UpdateFindingsFilter

See: newUpdateFindingsFilter smart constructor.

Constructors

UpdateFindingsFilter'

Fields

clientToken :: Maybe Text

A unique, case-sensitive token that you provide to ensure the idempotency of the request.

findingCriteria :: Maybe FindingCriteria

The criteria to use to filter findings.

action :: Maybe FindingsFilterAction

The action to perform on findings that meet the filter criteria (findingCriteria). Valid values are: ARCHIVE, suppress (automatically archive) the findings; and, NOOP, don't perform any action on the findings.

name :: Maybe Text

A custom name for the filter. The name must contain at least 3 characters and can contain as many as 64 characters.

We strongly recommend that you avoid including any sensitive data in the name of a filter. Other users might be able to see the filter's name, depending on the actions that they're allowed to perform in Amazon Macie.

description :: Maybe Text

A custom description of the filter. The description can contain as many as 512 characters.

We strongly recommend that you avoid including any sensitive data in the description of a filter. Other users might be able to see the filter's description, depending on the actions that they're allowed to perform in Amazon Macie.

position :: Maybe Int

The position of the filter in the list of saved filters on the Amazon Macie console. This value also determines the order in which the filter is applied to findings, relative to other filters that are also applied to the findings.

id :: Text

The unique identifier for the Amazon Macie resource or account that the request applies to.

newUpdateFindingsFilter

Arguments

:: Text
$sel:id:UpdateFindingsFilter'

-> UpdateFindingsFilter

Create a value of UpdateFindingsFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:UpdateFindingsFilter', updateFindingsFilter_clientToken - A unique, case-sensitive token that you provide to ensure the idempotency of the request.

$sel:findingCriteria:UpdateFindingsFilter', updateFindingsFilter_findingCriteria - The criteria to use to filter findings.

$sel:action:UpdateFindingsFilter', updateFindingsFilter_action - The action to perform on findings that meet the filter criteria (findingCriteria). Valid values are: ARCHIVE, suppress (automatically archive) the findings; and, NOOP, don't perform any action on the findings.

$sel:name:UpdateFindingsFilter', updateFindingsFilter_name - A custom name for the filter. The name must contain at least 3 characters and can contain as many as 64 characters.

We strongly recommend that you avoid including any sensitive data in the name of a filter. Other users might be able to see the filter's name, depending on the actions that they're allowed to perform in Amazon Macie.

$sel:description:UpdateFindingsFilter', updateFindingsFilter_description - A custom description of the filter. The description can contain as many as 512 characters.

We strongly recommend that you avoid including any sensitive data in the description of a filter. Other users might be able to see the filter's description, depending on the actions that they're allowed to perform in Amazon Macie.

$sel:position:UpdateFindingsFilter', updateFindingsFilter_position - The position of the filter in the list of saved filters on the Amazon Macie console. This value also determines the order in which the filter is applied to findings, relative to other filters that are also applied to the findings.

$sel:id:UpdateFindingsFilter', updateFindingsFilter_id - The unique identifier for the Amazon Macie resource or account that the request applies to.

Request Lenses

updateFindingsFilter_clientToken :: Lens' UpdateFindingsFilter (Maybe Text)

A unique, case-sensitive token that you provide to ensure the idempotency of the request.

updateFindingsFilter_findingCriteria :: Lens' UpdateFindingsFilter (Maybe FindingCriteria)

The criteria to use to filter findings.

updateFindingsFilter_action :: Lens' UpdateFindingsFilter (Maybe FindingsFilterAction)

The action to perform on findings that meet the filter criteria (findingCriteria). Valid values are: ARCHIVE, suppress (automatically archive) the findings; and, NOOP, don't perform any action on the findings.

updateFindingsFilter_name :: Lens' UpdateFindingsFilter (Maybe Text)

A custom name for the filter. The name must contain at least 3 characters and can contain as many as 64 characters.

We strongly recommend that you avoid including any sensitive data in the name of a filter. Other users might be able to see the filter's name, depending on the actions that they're allowed to perform in Amazon Macie.

updateFindingsFilter_description :: Lens' UpdateFindingsFilter (Maybe Text)

A custom description of the filter. The description can contain as many as 512 characters.

We strongly recommend that you avoid including any sensitive data in the description of a filter. Other users might be able to see the filter's description, depending on the actions that they're allowed to perform in Amazon Macie.

updateFindingsFilter_position :: Lens' UpdateFindingsFilter (Maybe Int)

The position of the filter in the list of saved filters on the Amazon Macie console. This value also determines the order in which the filter is applied to findings, relative to other filters that are also applied to the findings.

updateFindingsFilter_id :: Lens' UpdateFindingsFilter Text

The unique identifier for the Amazon Macie resource or account that the request applies to.

Destructuring the Response

data UpdateFindingsFilterResponse

See: newUpdateFindingsFilterResponse smart constructor.

Constructors

UpdateFindingsFilterResponse'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the filter that was updated.

id :: Maybe Text

The unique identifier for the filter that was updated.

httpStatus :: Int

The response's http status code.

newUpdateFindingsFilterResponse

Arguments

:: Int
$sel:httpStatus:UpdateFindingsFilterResponse'

-> UpdateFindingsFilterResponse

Create a value of UpdateFindingsFilterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:UpdateFindingsFilterResponse', updateFindingsFilterResponse_arn - The Amazon Resource Name (ARN) of the filter that was updated.

$sel:id:UpdateFindingsFilter', updateFindingsFilterResponse_id - The unique identifier for the filter that was updated.

$sel:httpStatus:UpdateFindingsFilterResponse', updateFindingsFilterResponse_httpStatus - The response's http status code.

Response Lenses

updateFindingsFilterResponse_arn :: Lens' UpdateFindingsFilterResponse (Maybe Text)

The Amazon Resource Name (ARN) of the filter that was updated.

updateFindingsFilterResponse_id :: Lens' UpdateFindingsFilterResponse (Maybe Text)

The unique identifier for the filter that was updated.

updateFindingsFilterResponse_httpStatus :: Lens' UpdateFindingsFilterResponse Int

The response's http status code.