Network.AWS.ManagedBlockChain.DeleteNode

Description

Deletes a node that your AWS account owns. All data on the node is lost and cannot be recovered.

Applies to Hyperledger Fabric and Ethereum.

Creating a Request

data DeleteNode

See: newDeleteNode smart constructor.

Constructors

DeleteNode'

Fields

memberId :: Maybe Text

The unique identifier of the member that owns this node.

Applies only to Hyperledger Fabric and is required for Hyperledger Fabric.

networkId :: Text

The unique identifier of the network that the node is on.

Ethereum public networks have the following NetworkIds:

n-ethereum-mainnet

n-ethereum-rinkeby

n-ethereum-ropsten

nodeId :: Text

The unique identifier of the node.

newDeleteNode

Arguments

:: Text
$sel:networkId:DeleteNode'

-> Text
$sel:nodeId:DeleteNode'

-> DeleteNode

Create a value of DeleteNode with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:memberId:DeleteNode', deleteNode_memberId - The unique identifier of the member that owns this node.

Applies only to Hyperledger Fabric and is required for Hyperledger Fabric.

$sel:networkId:DeleteNode', deleteNode_networkId - The unique identifier of the network that the node is on.

Ethereum public networks have the following NetworkIds:

n-ethereum-mainnet

n-ethereum-rinkeby

n-ethereum-ropsten

$sel:nodeId:DeleteNode', deleteNode_nodeId - The unique identifier of the node.

Request Lenses

deleteNode_memberId :: Lens' DeleteNode (Maybe Text)

The unique identifier of the member that owns this node.

Applies only to Hyperledger Fabric and is required for Hyperledger Fabric.

deleteNode_networkId :: Lens' DeleteNode Text

The unique identifier of the network that the node is on.

Ethereum public networks have the following NetworkIds:

n-ethereum-mainnet

n-ethereum-rinkeby

n-ethereum-ropsten

deleteNode_nodeId :: Lens' DeleteNode Text

The unique identifier of the node.

Destructuring the Response

data DeleteNodeResponse

See: newDeleteNodeResponse smart constructor.

Constructors

DeleteNodeResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteNodeResponse

Arguments

:: Int
$sel:httpStatus:DeleteNodeResponse'

-> DeleteNodeResponse

Create a value of DeleteNodeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteNodeResponse', deleteNodeResponse_httpStatus - The response's http status code.

Response Lenses

deleteNodeResponse_httpStatus :: Lens' DeleteNodeResponse Int

The response's http status code.