Network.AWS.ManagedBlockChain.GetMember

Description

Returns detailed information about a member.

Applies only to Hyperledger Fabric.

Creating a Request

data GetMember

See: newGetMember smart constructor.

Constructors

GetMember'

Fields

networkId :: Text

The unique identifier of the network to which the member belongs.

memberId :: Text

The unique identifier of the member.

newGetMember

Arguments

:: Text
$sel:networkId:GetMember'

-> Text
$sel:memberId:GetMember'

-> GetMember

Create a value of GetMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkId:GetMember', getMember_networkId - The unique identifier of the network to which the member belongs.

$sel:memberId:GetMember', getMember_memberId - The unique identifier of the member.

Request Lenses

getMember_networkId :: Lens' GetMember Text

The unique identifier of the network to which the member belongs.

getMember_memberId :: Lens' GetMember Text

The unique identifier of the member.

Destructuring the Response

data GetMemberResponse

See: newGetMemberResponse smart constructor.

Constructors

GetMemberResponse'

Fields

member :: Maybe Member

The properties of a member.

httpStatus :: Int

The response's http status code.

newGetMemberResponse

Arguments

:: Int
$sel:httpStatus:GetMemberResponse'

-> GetMemberResponse

Create a value of GetMemberResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:member:GetMemberResponse', getMemberResponse_member - The properties of a member.

$sel:httpStatus:GetMemberResponse', getMemberResponse_httpStatus - The response's http status code.

Response Lenses

getMemberResponse_member :: Lens' GetMemberResponse (Maybe Member)

The properties of a member.

getMemberResponse_httpStatus :: Lens' GetMemberResponse Int

The response's http status code.