Network.AWS.ManagedBlockChain.Lens

Description

Operations

CreateMember

createMember_clientRequestToken :: Lens' CreateMember Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the operation. An idempotent operation completes no more than one time. This identifier is required only if you make a service request directly using an HTTP client. It is generated automatically if you use an AWS SDK or the AWS CLI.

createMember_invitationId :: Lens' CreateMember Text

The unique identifier of the invitation that is sent to the member to join the network.

createMember_networkId :: Lens' CreateMember Text

The unique identifier of the network in which the member is created.

createMember_memberConfiguration :: Lens' CreateMember MemberConfiguration

Member configuration parameters.

createMemberResponse_memberId :: Lens' CreateMemberResponse (Maybe Text)

The unique identifier of the member.

createMemberResponse_httpStatus :: Lens' CreateMemberResponse Int

The response's http status code.

ListNetworks

listNetworks_status :: Lens' ListNetworks (Maybe NetworkStatus)

An optional status specifier. If provided, only networks currently in this status are listed.

Applies only to Hyperledger Fabric.

listNetworks_framework :: Lens' ListNetworks (Maybe Framework)

An optional framework specifier. If provided, only networks of this framework type are listed.

listNetworks_nextToken :: Lens' ListNetworks (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

listNetworks_name :: Lens' ListNetworks (Maybe Text)

The name of the network.

listNetworks_maxResults :: Lens' ListNetworks (Maybe Natural)

The maximum number of networks to list.

listNetworksResponse_networks :: Lens' ListNetworksResponse (Maybe [NetworkSummary])

An array of NetworkSummary objects that contain configuration properties for each network.

listNetworksResponse_nextToken :: Lens' ListNetworksResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

listNetworksResponse_httpStatus :: Lens' ListNetworksResponse Int

The response's http status code.

GetProposal

getProposal_networkId :: Lens' GetProposal Text

The unique identifier of the network for which the proposal is made.

getProposal_proposalId :: Lens' GetProposal Text

The unique identifier of the proposal.

getProposalResponse_proposal :: Lens' GetProposalResponse (Maybe Proposal)

Information about a proposal.

getProposalResponse_httpStatus :: Lens' GetProposalResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the resource. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The tags assigned to the resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

CreateNetwork

createNetwork_frameworkConfiguration :: Lens' CreateNetwork (Maybe NetworkFrameworkConfiguration)

Configuration properties of the blockchain framework relevant to the network configuration.

createNetwork_description :: Lens' CreateNetwork (Maybe Text)

An optional description for the network.

createNetwork_tags :: Lens' CreateNetwork (Maybe (HashMap Text Text))

Tags to assign to the network. Each tag consists of a key and optional value.

When specifying tags during creation, you can specify multiple key-value pairs in a single request, with an overall maximum of 50 tags added to each resource.

For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Ethereum Developer Guide, or Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

createNetwork_clientRequestToken :: Lens' CreateNetwork Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the operation. An idempotent operation completes no more than one time. This identifier is required only if you make a service request directly using an HTTP client. It is generated automatically if you use an AWS SDK or the AWS CLI.

createNetwork_name :: Lens' CreateNetwork Text

The name of the network.

createNetwork_framework :: Lens' CreateNetwork Framework

The blockchain framework that the network uses.

createNetwork_frameworkVersion :: Lens' CreateNetwork Text

The version of the blockchain framework that the network uses.

createNetwork_votingPolicy :: Lens' CreateNetwork VotingPolicy

The voting rules used by the network to determine if a proposal is approved.

createNetwork_memberConfiguration :: Lens' CreateNetwork MemberConfiguration

Configuration properties for the first member within the network.

createNetworkResponse_memberId :: Lens' CreateNetworkResponse (Maybe Text)

The unique identifier for the first member within the network.

createNetworkResponse_networkId :: Lens' CreateNetworkResponse (Maybe Text)

The unique identifier for the network.

createNetworkResponse_httpStatus :: Lens' CreateNetworkResponse Int

The response's http status code.

ListProposals

listProposals_nextToken :: Lens' ListProposals (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

listProposals_maxResults :: Lens' ListProposals (Maybe Natural)

The maximum number of proposals to return.

listProposals_networkId :: Lens' ListProposals Text

The unique identifier of the network.

listProposalsResponse_proposals :: Lens' ListProposalsResponse (Maybe [ProposalSummary])

The summary of each proposal made on the network.

listProposalsResponse_nextToken :: Lens' ListProposalsResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

listProposalsResponse_httpStatus :: Lens' ListProposalsResponse Int

The response's http status code.

ListInvitations

listInvitations_nextToken :: Lens' ListInvitations (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

listInvitations_maxResults :: Lens' ListInvitations (Maybe Natural)

The maximum number of invitations to return.

listInvitationsResponse_invitations :: Lens' ListInvitationsResponse (Maybe [Invitation])

The invitations for the network.

listInvitationsResponse_nextToken :: Lens' ListInvitationsResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

listInvitationsResponse_httpStatus :: Lens' ListInvitationsResponse Int

The response's http status code.

CreateProposal

createProposal_description :: Lens' CreateProposal (Maybe Text)

A description for the proposal that is visible to voting members, for example, "Proposal to add Example Corp. as member."

createProposal_tags :: Lens' CreateProposal (Maybe (HashMap Text Text))

Tags to assign to the proposal. Each tag consists of a key and optional value.

When specifying tags during creation, you can specify multiple key-value pairs in a single request, with an overall maximum of 50 tags added to each resource. If the proposal is for a network invitation, the invitation inherits the tags added to the proposal.

For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Ethereum Developer Guide, or Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

createProposal_clientRequestToken :: Lens' CreateProposal Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the operation. An idempotent operation completes no more than one time. This identifier is required only if you make a service request directly using an HTTP client. It is generated automatically if you use an AWS SDK or the AWS CLI.

createProposal_networkId :: Lens' CreateProposal Text

The unique identifier of the network for which the proposal is made.

createProposal_memberId :: Lens' CreateProposal Text

The unique identifier of the member that is creating the proposal. This identifier is especially useful for identifying the member making the proposal when multiple members exist in a single AWS account.

createProposal_actions :: Lens' CreateProposal ProposalActions

The type of actions proposed, such as inviting a member or removing a member. The types of Actions in a proposal are mutually exclusive. For example, a proposal with Invitations actions cannot also contain Removals actions.

createProposalResponse_proposalId :: Lens' CreateProposalResponse (Maybe Text)

The unique identifier of the proposal.

createProposalResponse_httpStatus :: Lens' CreateProposalResponse Int

The response's http status code.

GetNetwork

getNetwork_networkId :: Lens' GetNetwork Text

The unique identifier of the network to get information about.

getNetworkResponse_network :: Lens' GetNetworkResponse (Maybe Network)

An object containing network configuration parameters.

getNetworkResponse_httpStatus :: Lens' GetNetworkResponse Int

The response's http status code.

GetNode

getNode_memberId :: Lens' GetNode (Maybe Text)

The unique identifier of the member that owns the node.

Applies only to Hyperledger Fabric and is required for Hyperledger Fabric.

getNode_networkId :: Lens' GetNode Text

The unique identifier of the network that the node is on.

getNode_nodeId :: Lens' GetNode Text

The unique identifier of the node.

getNodeResponse_node :: Lens' GetNodeResponse (Maybe Node)

Properties of the node configuration.

getNodeResponse_httpStatus :: Lens' GetNodeResponse Int

The response's http status code.

UpdateMember

updateMember_logPublishingConfiguration :: Lens' UpdateMember (Maybe MemberLogPublishingConfiguration)

Configuration properties for publishing to Amazon CloudWatch Logs.

updateMember_networkId :: Lens' UpdateMember Text

The unique identifier of the Managed Blockchain network to which the member belongs.

updateMember_memberId :: Lens' UpdateMember Text

The unique identifier of the member.

updateMemberResponse_httpStatus :: Lens' UpdateMemberResponse Int

The response's http status code.

DeleteMember

deleteMember_networkId :: Lens' DeleteMember Text

The unique identifier of the network from which the member is removed.

deleteMember_memberId :: Lens' DeleteMember Text

The unique identifier of the member to remove.

deleteMemberResponse_httpStatus :: Lens' DeleteMemberResponse Int

The response's http status code.

ListMembers

listMembers_status :: Lens' ListMembers (Maybe MemberStatus)

An optional status specifier. If provided, only members currently in this status are listed.

listMembers_nextToken :: Lens' ListMembers (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

listMembers_name :: Lens' ListMembers (Maybe Text)

The optional name of the member to list.

listMembers_isOwned :: Lens' ListMembers (Maybe Bool)

An optional Boolean value. If provided, the request is limited either to members that the current AWS account owns (true) or that other AWS accounts own (false). If omitted, all members are listed.

listMembers_maxResults :: Lens' ListMembers (Maybe Natural)

The maximum number of members to return in the request.

listMembers_networkId :: Lens' ListMembers Text

The unique identifier of the network for which to list members.

listMembersResponse_members :: Lens' ListMembersResponse (Maybe [MemberSummary])

An array of MemberSummary objects. Each object contains details about a network member.

listMembersResponse_nextToken :: Lens' ListMembersResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

listMembersResponse_httpStatus :: Lens' ListMembersResponse Int

The response's http status code.

CreateNode

createNode_memberId :: Lens' CreateNode (Maybe Text)

The unique identifier of the member that owns this node.

Applies only to Hyperledger Fabric.

createNode_tags :: Lens' CreateNode (Maybe (HashMap Text Text))

Tags to assign to the node. Each tag consists of a key and optional value.

When specifying tags during creation, you can specify multiple key-value pairs in a single request, with an overall maximum of 50 tags added to each resource.

For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Ethereum Developer Guide, or Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

createNode_clientRequestToken :: Lens' CreateNode Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the operation. An idempotent operation completes no more than one time. This identifier is required only if you make a service request directly using an HTTP client. It is generated automatically if you use an AWS SDK or the AWS CLI.

createNode_networkId :: Lens' CreateNode Text

The unique identifier of the network for the node.

Ethereum public networks have the following NetworkIds:

n-ethereum-mainnet

n-ethereum-rinkeby

n-ethereum-ropsten

createNode_nodeConfiguration :: Lens' CreateNode NodeConfiguration

The properties of a node configuration.

createNodeResponse_nodeId :: Lens' CreateNodeResponse (Maybe Text)

The unique identifier of the node.

createNodeResponse_httpStatus :: Lens' CreateNodeResponse Int

The response's http status code.

ListProposalVotes

listProposalVotes_nextToken :: Lens' ListProposalVotes (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

listProposalVotes_maxResults :: Lens' ListProposalVotes (Maybe Natural)

The maximum number of votes to return.

listProposalVotes_networkId :: Lens' ListProposalVotes Text

The unique identifier of the network.

listProposalVotes_proposalId :: Lens' ListProposalVotes Text

The unique identifier of the proposal.

listProposalVotesResponse_nextToken :: Lens' ListProposalVotesResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

listProposalVotesResponse_proposalVotes :: Lens' ListProposalVotesResponse (Maybe [VoteSummary])

The list of votes.

listProposalVotesResponse_httpStatus :: Lens' ListProposalVotesResponse Int

The response's http status code.

VoteOnProposal

voteOnProposal_networkId :: Lens' VoteOnProposal Text

The unique identifier of the network.

voteOnProposal_proposalId :: Lens' VoteOnProposal Text

The unique identifier of the proposal.

voteOnProposal_voterMemberId :: Lens' VoteOnProposal Text

The unique identifier of the member casting the vote.

voteOnProposal_vote :: Lens' VoteOnProposal VoteValue

The value of the vote.

voteOnProposalResponse_httpStatus :: Lens' VoteOnProposalResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the resource. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The tags to assign to the specified resource. Tag values can be empty, for example, "MyTagKey" : "". You can specify multiple key-value pairs in a single request, with an overall maximum of 50 tags added to each resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the resource. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

untagResource_tagKeys :: Lens' UntagResource [Text]

The tag keys.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

GetMember

getMember_networkId :: Lens' GetMember Text

The unique identifier of the network to which the member belongs.

getMember_memberId :: Lens' GetMember Text

The unique identifier of the member.

getMemberResponse_member :: Lens' GetMemberResponse (Maybe Member)

The properties of a member.

getMemberResponse_httpStatus :: Lens' GetMemberResponse Int

The response's http status code.

RejectInvitation

rejectInvitation_invitationId :: Lens' RejectInvitation Text

The unique identifier of the invitation to reject.

rejectInvitationResponse_httpStatus :: Lens' RejectInvitationResponse Int

The response's http status code.

UpdateNode

updateNode_logPublishingConfiguration :: Lens' UpdateNode (Maybe NodeLogPublishingConfiguration)

Configuration properties for publishing to Amazon CloudWatch Logs.

updateNode_memberId :: Lens' UpdateNode (Maybe Text)

The unique identifier of the member that owns the node.

Applies only to Hyperledger Fabric.

updateNode_networkId :: Lens' UpdateNode Text

The unique identifier of the network that the node is on.

updateNode_nodeId :: Lens' UpdateNode Text

The unique identifier of the node.

updateNodeResponse_httpStatus :: Lens' UpdateNodeResponse Int

The response's http status code.

DeleteNode

deleteNode_memberId :: Lens' DeleteNode (Maybe Text)

The unique identifier of the member that owns this node.

Applies only to Hyperledger Fabric and is required for Hyperledger Fabric.

deleteNode_networkId :: Lens' DeleteNode Text

The unique identifier of the network that the node is on.

Ethereum public networks have the following NetworkIds:

n-ethereum-mainnet

n-ethereum-rinkeby

n-ethereum-ropsten

deleteNode_nodeId :: Lens' DeleteNode Text

The unique identifier of the node.

deleteNodeResponse_httpStatus :: Lens' DeleteNodeResponse Int

The response's http status code.

ListNodes

listNodes_status :: Lens' ListNodes (Maybe NodeStatus)

An optional status specifier. If provided, only nodes currently in this status are listed.

listNodes_memberId :: Lens' ListNodes (Maybe Text)

The unique identifier of the member who owns the nodes to list.

Applies only to Hyperledger Fabric and is required for Hyperledger Fabric.

listNodes_nextToken :: Lens' ListNodes (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

listNodes_maxResults :: Lens' ListNodes (Maybe Natural)

The maximum number of nodes to list.

listNodes_networkId :: Lens' ListNodes Text

The unique identifier of the network for which to list nodes.

listNodesResponse_nextToken :: Lens' ListNodesResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

listNodesResponse_nodes :: Lens' ListNodesResponse (Maybe [NodeSummary])

An array of NodeSummary objects that contain configuration properties for each node.

listNodesResponse_httpStatus :: Lens' ListNodesResponse Int

The response's http status code.

Types

ApprovalThresholdPolicy

approvalThresholdPolicy_thresholdPercentage :: Lens' ApprovalThresholdPolicy (Maybe Natural)

The percentage of votes among all members that must be YES for a proposal to be approved. For example, a ThresholdPercentage value of 50 indicates 50%. The ThresholdComparator determines the precise comparison. If a ThresholdPercentage value of 50 is specified on a network with 10 members, along with a ThresholdComparator value of GREATER_THAN, this indicates that 6 YES votes are required for the proposal to be approved.

approvalThresholdPolicy_thresholdComparator :: Lens' ApprovalThresholdPolicy (Maybe ThresholdComparator)

Determines whether the vote percentage must be greater than the ThresholdPercentage or must be greater than or equal to the ThreholdPercentage to be approved.

approvalThresholdPolicy_proposalDurationInHours :: Lens' ApprovalThresholdPolicy (Maybe Natural)

The duration from the time that a proposal is created until it expires. If members cast neither the required number of YES votes to approve the proposal nor the number of NO votes required to reject it before the duration expires, the proposal is EXPIRED and ProposalActions are not carried out.

Invitation

invitation_status :: Lens' Invitation (Maybe InvitationStatus)

The status of the invitation:

PENDING - The invitee has not created a member to join the network, and the invitation has not yet expired.

ACCEPTING - The invitee has begun creating a member, and creation has not yet completed.

ACCEPTED - The invitee created a member and joined the network using the InvitationID.

REJECTED - The invitee rejected the invitation.

EXPIRED - The invitee neither created a member nor rejected the invitation before the ExpirationDate.

invitation_arn :: Lens' Invitation (Maybe Text)

The Amazon Resource Name (ARN) of the invitation. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

invitation_invitationId :: Lens' Invitation (Maybe Text)

The unique identifier for the invitation.

invitation_expirationDate :: Lens' Invitation (Maybe UTCTime)

The date and time that the invitation expires. This is the CreationDate plus the ProposalDurationInHours that is specified in the ProposalThresholdPolicy. After this date and time, the invitee can no longer create a member and join the network using this InvitationId.

invitation_networkSummary :: Lens' Invitation (Maybe NetworkSummary)

Undocumented member.

invitation_creationDate :: Lens' Invitation (Maybe UTCTime)

The date and time that the invitation was created.

InviteAction

inviteAction_principal :: Lens' InviteAction Text

The AWS account ID to invite.

LogConfiguration

logConfiguration_enabled :: Lens' LogConfiguration (Maybe Bool)

Indicates whether logging is enabled.

LogConfigurations

logConfigurations_cloudwatch :: Lens' LogConfigurations (Maybe LogConfiguration)

Parameters for publishing logs to Amazon CloudWatch Logs.

Member

member_status :: Lens' Member (Maybe MemberStatus)

The status of a member.

CREATING - The AWS account is in the process of creating a member.

AVAILABLE - The member has been created and can participate in the network.

CREATE_FAILED - The AWS account attempted to create a member and creation failed.

UPDATING - The member is in the process of being updated.

DELETING - The member and all associated resources are in the process of being deleted. Either the AWS account that owns the member deleted it, or the member is being deleted as the result of an APPROVED PROPOSAL to remove the member.

DELETED - The member can no longer participate on the network and all associated resources are deleted. Either the AWS account that owns the member deleted it, or the member is being deleted as the result of an APPROVED PROPOSAL to remove the member.

INACCESSIBLE_ENCRYPTION_KEY - The member is impaired and might not function as expected because it cannot access the specified customer managed key in AWS KMS for encryption at rest. Either the KMS key was disabled or deleted, or the grants on the key were revoked.

The effect of disabling or deleting a key, or revoking a grant is not immediate. The member resource might take some time to find that the key is inaccessible. When a resource is in this state, we recommend deleting and recreating the resource.

member_kmsKeyArn :: Lens' Member (Maybe Text)

The Amazon Resource Name (ARN) of the customer managed key in AWS Key Management Service (AWS KMS) that the member uses for encryption at rest. If the value of this parameter is "AWS Owned KMS Key", the member uses an AWS owned KMS key for encryption. This parameter is inherited by the nodes that this member owns.

member_logPublishingConfiguration :: Lens' Member (Maybe MemberLogPublishingConfiguration)

Configuration properties for logging events associated with a member.

member_arn :: Lens' Member (Maybe Text)

The Amazon Resource Name (ARN) of the member. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

member_networkId :: Lens' Member (Maybe Text)

The unique identifier of the network to which the member belongs.

member_name :: Lens' Member (Maybe Text)

The name of the member.

member_id :: Lens' Member (Maybe Text)

The unique identifier of the member.

member_creationDate :: Lens' Member (Maybe UTCTime)

The date and time that the member was created.

member_frameworkAttributes :: Lens' Member (Maybe MemberFrameworkAttributes)

Attributes relevant to a member for the blockchain framework that the Managed Blockchain network uses.

member_description :: Lens' Member (Maybe Text)

An optional description for the member.

member_tags :: Lens' Member (Maybe (HashMap Text Text))

Tags assigned to the member. Tags consist of a key and optional value. For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

MemberConfiguration

memberConfiguration_kmsKeyArn :: Lens' MemberConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the customer managed key in AWS Key Management Service (AWS KMS) to use for encryption at rest in the member. This parameter is inherited by any nodes that this member creates.

Use one of the following options to specify this parameter:

Undefined or empty string - The member uses an AWS owned KMS key for encryption by default.

A valid symmetric customer managed KMS key - The member uses the specified key for encryption.

Amazon Managed Blockchain doesn't support asymmetric keys. For more information, see Using symmetric and asymmetric keys in the AWS Key Management Service Developer Guide.

The following is an example of a KMS key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

memberConfiguration_logPublishingConfiguration :: Lens' MemberConfiguration (Maybe MemberLogPublishingConfiguration)

Configuration properties for logging events associated with a member of a Managed Blockchain network.

memberConfiguration_description :: Lens' MemberConfiguration (Maybe Text)

An optional description of the member.

memberConfiguration_tags :: Lens' MemberConfiguration (Maybe (HashMap Text Text))

Tags assigned to the member. Tags consist of a key and optional value. For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

When specifying tags during creation, you can specify multiple key-value pairs in a single request, with an overall maximum of 50 tags added to each resource.

memberConfiguration_name :: Lens' MemberConfiguration Text

The name of the member.

memberConfiguration_frameworkConfiguration :: Lens' MemberConfiguration MemberFrameworkConfiguration

Configuration properties of the blockchain framework relevant to the member.

MemberFabricAttributes

memberFabricAttributes_caEndpoint :: Lens' MemberFabricAttributes (Maybe Text)

The endpoint used to access the member's certificate authority.

memberFabricAttributes_adminUsername :: Lens' MemberFabricAttributes (Maybe Text)

The user name for the initial administrator user for the member.

MemberFabricConfiguration

memberFabricConfiguration_adminUsername :: Lens' MemberFabricConfiguration Text

The user name for the member's initial administrative user.

memberFabricConfiguration_adminPassword :: Lens' MemberFabricConfiguration Text

The password for the member's initial administrative user. The AdminPassword must be at least eight characters long and no more than 32 characters. It must contain at least one uppercase letter, one lowercase letter, and one digit. It cannot have a single quotation mark (‘), a double quotation marks (“), a forward slash(/), a backward slash(\), @, or a space.

MemberFabricLogPublishingConfiguration

memberFabricLogPublishingConfiguration_caLogs :: Lens' MemberFabricLogPublishingConfiguration (Maybe LogConfigurations)

Configuration properties for logging events associated with a member's Certificate Authority (CA). CA logs help you determine when a member in your account joins the network, or when new peers register with a member CA.

MemberFrameworkAttributes

memberFrameworkAttributes_fabric :: Lens' MemberFrameworkAttributes (Maybe MemberFabricAttributes)

Attributes of Hyperledger Fabric relevant to a member on a Managed Blockchain network that uses Hyperledger Fabric.

MemberFrameworkConfiguration

memberFrameworkConfiguration_fabric :: Lens' MemberFrameworkConfiguration (Maybe MemberFabricConfiguration)

Attributes of Hyperledger Fabric for a member on a Managed Blockchain network that uses Hyperledger Fabric.

MemberLogPublishingConfiguration

memberLogPublishingConfiguration_fabric :: Lens' MemberLogPublishingConfiguration (Maybe MemberFabricLogPublishingConfiguration)

Configuration properties for logging events associated with a member of a Managed Blockchain network using the Hyperledger Fabric framework.

MemberSummary

memberSummary_status :: Lens' MemberSummary (Maybe MemberStatus)

The status of the member.

CREATING - The AWS account is in the process of creating a member.

AVAILABLE - The member has been created and can participate in the network.

CREATE_FAILED - The AWS account attempted to create a member and creation failed.

UPDATING - The member is in the process of being updated.

DELETING - The member and all associated resources are in the process of being deleted. Either the AWS account that owns the member deleted it, or the member is being deleted as the result of an APPROVED PROPOSAL to remove the member.

DELETED - The member can no longer participate on the network and all associated resources are deleted. Either the AWS account that owns the member deleted it, or the member is being deleted as the result of an APPROVED PROPOSAL to remove the member.

INACCESSIBLE_ENCRYPTION_KEY - The member is impaired and might not function as expected because it cannot access the specified customer managed key in AWS Key Management Service (AWS KMS) for encryption at rest. Either the KMS key was disabled or deleted, or the grants on the key were revoked.

The effect of disabling or deleting a key, or revoking a grant is not immediate. The member resource might take some time to find that the key is inaccessible. When a resource is in this state, we recommend deleting and recreating the resource.

memberSummary_arn :: Lens' MemberSummary (Maybe Text)

The Amazon Resource Name (ARN) of the member. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

memberSummary_name :: Lens' MemberSummary (Maybe Text)

The name of the member.

memberSummary_id :: Lens' MemberSummary (Maybe Text)

The unique identifier of the member.

memberSummary_isOwned :: Lens' MemberSummary (Maybe Bool)

An indicator of whether the member is owned by your AWS account or a different AWS account.

memberSummary_creationDate :: Lens' MemberSummary (Maybe UTCTime)

The date and time that the member was created.

memberSummary_description :: Lens' MemberSummary (Maybe Text)

An optional description of the member.

Network

network_status :: Lens' Network (Maybe NetworkStatus)

The current status of the network.

network_framework :: Lens' Network (Maybe Framework)

The blockchain framework that the network uses.

network_arn :: Lens' Network (Maybe Text)

The Amazon Resource Name (ARN) of the network. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

network_frameworkVersion :: Lens' Network (Maybe Text)

The version of the blockchain framework that the network uses.

network_vpcEndpointServiceName :: Lens' Network (Maybe Text)

The VPC endpoint service name of the VPC endpoint service of the network. Members use the VPC endpoint service name to create a VPC endpoint to access network resources.

network_name :: Lens' Network (Maybe Text)

The name of the network.

network_id :: Lens' Network (Maybe Text)

The unique identifier of the network.

network_votingPolicy :: Lens' Network (Maybe VotingPolicy)

The voting rules for the network to decide if a proposal is accepted.

network_creationDate :: Lens' Network (Maybe UTCTime)

The date and time that the network was created.

network_frameworkAttributes :: Lens' Network (Maybe NetworkFrameworkAttributes)

Attributes of the blockchain framework that the network uses.

network_description :: Lens' Network (Maybe Text)

Attributes of the blockchain framework for the network.

network_tags :: Lens' Network (Maybe (HashMap Text Text))

Tags assigned to the network. Each tag consists of a key and optional value.

For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Ethereum Developer Guide, or Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

NetworkEthereumAttributes

networkEthereumAttributes_chainId :: Lens' NetworkEthereumAttributes (Maybe Text)

The Ethereum CHAIN_ID associated with the Ethereum network. Chain IDs are as follows:

mainnet = 1

rinkeby = 4

ropsten = 3

NetworkFabricAttributes

networkFabricAttributes_edition :: Lens' NetworkFabricAttributes (Maybe Edition)

The edition of Amazon Managed Blockchain that Hyperledger Fabric uses. For more information, see Amazon Managed Blockchain Pricing.

networkFabricAttributes_orderingServiceEndpoint :: Lens' NetworkFabricAttributes (Maybe Text)

The endpoint of the ordering service for the network.

NetworkFabricConfiguration

networkFabricConfiguration_edition :: Lens' NetworkFabricConfiguration Edition

The edition of Amazon Managed Blockchain that the network uses. For more information, see Amazon Managed Blockchain Pricing.

NetworkFrameworkAttributes

networkFrameworkAttributes_fabric :: Lens' NetworkFrameworkAttributes (Maybe NetworkFabricAttributes)

Attributes of Hyperledger Fabric for a Managed Blockchain network that uses Hyperledger Fabric.

networkFrameworkAttributes_ethereum :: Lens' NetworkFrameworkAttributes (Maybe NetworkEthereumAttributes)

Attributes of an Ethereum network for Managed Blockchain resources participating in an Ethereum network.

NetworkFrameworkConfiguration

networkFrameworkConfiguration_fabric :: Lens' NetworkFrameworkConfiguration (Maybe NetworkFabricConfiguration)

Hyperledger Fabric configuration properties for a Managed Blockchain network that uses Hyperledger Fabric.

NetworkSummary

networkSummary_status :: Lens' NetworkSummary (Maybe NetworkStatus)

The current status of the network.

networkSummary_framework :: Lens' NetworkSummary (Maybe Framework)

The blockchain framework that the network uses.

networkSummary_arn :: Lens' NetworkSummary (Maybe Text)

The Amazon Resource Name (ARN) of the network. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

networkSummary_frameworkVersion :: Lens' NetworkSummary (Maybe Text)

The version of the blockchain framework that the network uses.

networkSummary_name :: Lens' NetworkSummary (Maybe Text)

The name of the network.

networkSummary_id :: Lens' NetworkSummary (Maybe Text)

The unique identifier of the network.

networkSummary_creationDate :: Lens' NetworkSummary (Maybe UTCTime)

The date and time that the network was created.

networkSummary_description :: Lens' NetworkSummary (Maybe Text)

An optional description of the network.

Node

node_status :: Lens' Node (Maybe NodeStatus)

The status of the node.

CREATING - The AWS account is in the process of creating a node.

AVAILABLE - The node has been created and can participate in the network.

UNHEALTHY - The node is impaired and might not function as expected. Amazon Managed Blockchain automatically finds nodes in this state and tries to recover them. If a node is recoverable, it returns to AVAILABLE. Otherwise, it moves to FAILED status.

CREATE_FAILED - The AWS account attempted to create a node and creation failed.

UPDATING - The node is in the process of being updated.

DELETING - The node is in the process of being deleted.

DELETED - The node can no longer participate on the network.

FAILED - The node is no longer functional, cannot be recovered, and must be deleted.

INACCESSIBLE_ENCRYPTION_KEY - The node is impaired and might not function as expected because it cannot access the specified customer managed key in AWS KMS for encryption at rest. Either the KMS key was disabled or deleted, or the grants on the key were revoked.

The effect of disabling or deleting a key, or revoking a grant is not immediate. The node resource might take some time to find that the key is inaccessible. When a resource is in this state, we recommend deleting and recreating the resource.

node_kmsKeyArn :: Lens' Node (Maybe Text)

The Amazon Resource Name (ARN) of the customer managed key in AWS Key Management Service (AWS KMS) that the node uses for encryption at rest. If the value of this parameter is "AWS Owned KMS Key", the node uses an AWS owned KMS key for encryption. The node inherits this parameter from the member that it belongs to.

Applies only to Hyperledger Fabric.

node_logPublishingConfiguration :: Lens' Node (Maybe NodeLogPublishingConfiguration)

Configuration properties for logging events associated with a peer node on a Hyperledger Fabric network on Managed Blockchain.

node_memberId :: Lens' Node (Maybe Text)

The unique identifier of the member to which the node belongs.

Applies only to Hyperledger Fabric.

node_arn :: Lens' Node (Maybe Text)

The Amazon Resource Name (ARN) of the node. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

node_networkId :: Lens' Node (Maybe Text)

The unique identifier of the network that the node is on.

node_instanceType :: Lens' Node (Maybe Text)

The instance type of the node.

node_stateDB :: Lens' Node (Maybe StateDBType)

The state database that the node uses. Values are LevelDB or CouchDB.

Applies only to Hyperledger Fabric.

node_availabilityZone :: Lens' Node (Maybe Text)

The Availability Zone in which the node exists. Required for Ethereum nodes.

node_id :: Lens' Node (Maybe Text)

The unique identifier of the node.

node_creationDate :: Lens' Node (Maybe UTCTime)

The date and time that the node was created.

node_frameworkAttributes :: Lens' Node (Maybe NodeFrameworkAttributes)

Attributes of the blockchain framework being used.

node_tags :: Lens' Node (Maybe (HashMap Text Text))

Tags assigned to the node. Each tag consists of a key and optional value.

For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Ethereum Developer Guide, or Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

NodeConfiguration

nodeConfiguration_logPublishingConfiguration :: Lens' NodeConfiguration (Maybe NodeLogPublishingConfiguration)

Configuration properties for logging events associated with a peer node on a Hyperledger Fabric network on Managed Blockchain.

nodeConfiguration_stateDB :: Lens' NodeConfiguration (Maybe StateDBType)

The state database that the node uses. Values are LevelDB or CouchDB. When using an Amazon Managed Blockchain network with Hyperledger Fabric version 1.4 or later, the default is CouchDB.

Applies only to Hyperledger Fabric.

nodeConfiguration_availabilityZone :: Lens' NodeConfiguration (Maybe Text)

The Availability Zone in which the node exists. Required for Ethereum nodes.

nodeConfiguration_instanceType :: Lens' NodeConfiguration Text

The Amazon Managed Blockchain instance type for the node.

NodeEthereumAttributes

nodeEthereumAttributes_httpEndpoint :: Lens' NodeEthereumAttributes (Maybe Text)

The endpoint on which the Ethereum node listens to run Ethereum JSON-RPC methods over HTTP connections from a client. Use this endpoint in client code for smart contracts when using an HTTP connection. Connections to this endpoint are authenticated using Signature Version 4.

nodeEthereumAttributes_webSocketEndpoint :: Lens' NodeEthereumAttributes (Maybe Text)

The endpoint on which the Ethereum node listens to run Ethereum JSON-RPC methods over WebSockets connections from a client. Use this endpoint in client code for smart contracts when using a WebSockets connection. Connections to this endpoint are authenticated using Signature Version 4.

NodeFabricAttributes

nodeFabricAttributes_peerEventEndpoint :: Lens' NodeFabricAttributes (Maybe Text)

The endpoint that identifies the peer node for peer channel-based event services.

nodeFabricAttributes_peerEndpoint :: Lens' NodeFabricAttributes (Maybe Text)

The endpoint that identifies the peer node for all services except peer channel-based event services.

NodeFabricLogPublishingConfiguration

nodeFabricLogPublishingConfiguration_chaincodeLogs :: Lens' NodeFabricLogPublishingConfiguration (Maybe LogConfigurations)

Configuration properties for logging events associated with chaincode execution on a peer node. Chaincode logs contain the results of instantiating, invoking, and querying the chaincode. A peer can run multiple instances of chaincode. When enabled, a log stream is created for all chaincodes, with an individual log stream for each chaincode.

nodeFabricLogPublishingConfiguration_peerLogs :: Lens' NodeFabricLogPublishingConfiguration (Maybe LogConfigurations)

Configuration properties for a peer node log. Peer node logs contain messages generated when your client submits transaction proposals to peer nodes, requests to join channels, enrolls an admin peer, and lists the chaincode instances on a peer node.

NodeFrameworkAttributes

nodeFrameworkAttributes_fabric :: Lens' NodeFrameworkAttributes (Maybe NodeFabricAttributes)

Attributes of Hyperledger Fabric for a peer node on a Managed Blockchain network that uses Hyperledger Fabric.

nodeFrameworkAttributes_ethereum :: Lens' NodeFrameworkAttributes (Maybe NodeEthereumAttributes)

Attributes of Ethereum for a node on a Managed Blockchain network that uses Ethereum.

NodeLogPublishingConfiguration

nodeLogPublishingConfiguration_fabric :: Lens' NodeLogPublishingConfiguration (Maybe NodeFabricLogPublishingConfiguration)

Configuration properties for logging events associated with a node that is owned by a member of a Managed Blockchain network using the Hyperledger Fabric framework.

NodeSummary

nodeSummary_status :: Lens' NodeSummary (Maybe NodeStatus)

The status of the node.

nodeSummary_arn :: Lens' NodeSummary (Maybe Text)

The Amazon Resource Name (ARN) of the node. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

nodeSummary_instanceType :: Lens' NodeSummary (Maybe Text)

The EC2 instance type for the node.

nodeSummary_availabilityZone :: Lens' NodeSummary (Maybe Text)

The Availability Zone in which the node exists.

nodeSummary_id :: Lens' NodeSummary (Maybe Text)

The unique identifier of the node.

nodeSummary_creationDate :: Lens' NodeSummary (Maybe UTCTime)

The date and time that the node was created.

Proposal

proposal_status :: Lens' Proposal (Maybe ProposalStatus)

The status of the proposal. Values are as follows:

IN_PROGRESS - The proposal is active and open for member voting.

APPROVED - The proposal was approved with sufficient YES votes among members according to the VotingPolicy specified for the Network. The specified proposal actions are carried out.

REJECTED - The proposal was rejected with insufficient YES votes among members according to the VotingPolicy specified for the Network. The specified ProposalActions are not carried out.

EXPIRED - Members did not cast the number of votes required to determine the proposal outcome before the proposal expired. The specified ProposalActions are not carried out.

ACTION_FAILED - One or more of the specified ProposalActions in a proposal that was approved could not be completed because of an error. The ACTION_FAILED status occurs even if only one ProposalAction fails and other actions are successful.

proposal_yesVoteCount :: Lens' Proposal (Maybe Int)

The current total of YES votes cast on the proposal by members.

proposal_noVoteCount :: Lens' Proposal (Maybe Int)

The current total of NO votes cast on the proposal by members.

proposal_arn :: Lens' Proposal (Maybe Text)

The Amazon Resource Name (ARN) of the proposal. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

proposal_actions :: Lens' Proposal (Maybe ProposalActions)

The actions to perform on the network if the proposal is APPROVED.

proposal_networkId :: Lens' Proposal (Maybe Text)

The unique identifier of the network for which the proposal is made.

proposal_proposedByMemberId :: Lens' Proposal (Maybe Text)

The unique identifier of the member that created the proposal.

proposal_proposalId :: Lens' Proposal (Maybe Text)

The unique identifier of the proposal.

proposal_proposedByMemberName :: Lens' Proposal (Maybe Text)

The name of the member that created the proposal.

proposal_expirationDate :: Lens' Proposal (Maybe UTCTime)

The date and time that the proposal expires. This is the CreationDate plus the ProposalDurationInHours that is specified in the ProposalThresholdPolicy. After this date and time, if members have not cast enough votes to determine the outcome according to the voting policy, the proposal is EXPIRED and Actions are not carried out.

proposal_creationDate :: Lens' Proposal (Maybe UTCTime)

The date and time that the proposal was created.

proposal_description :: Lens' Proposal (Maybe Text)

The description of the proposal.

proposal_tags :: Lens' Proposal (Maybe (HashMap Text Text))

Tags assigned to the proposal. Each tag consists of a key and optional value.

For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Ethereum Developer Guide, or Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

proposal_outstandingVoteCount :: Lens' Proposal (Maybe Int)

The number of votes remaining to be cast on the proposal by members. In other words, the number of members minus the sum of YES votes and NO votes.

ProposalActions

proposalActions_invitations :: Lens' ProposalActions (Maybe [InviteAction])

The actions to perform for an APPROVED proposal to invite an AWS account to create a member and join the network.

proposalActions_removals :: Lens' ProposalActions (Maybe [RemoveAction])

The actions to perform for an APPROVED proposal to remove a member from the network, which deletes the member and all associated member resources from the network.

ProposalSummary

proposalSummary_status :: Lens' ProposalSummary (Maybe ProposalStatus)

The status of the proposal. Values are as follows:

IN_PROGRESS - The proposal is active and open for member voting.

APPROVED - The proposal was approved with sufficient YES votes among members according to the VotingPolicy specified for the Network. The specified proposal actions are carried out.

REJECTED - The proposal was rejected with insufficient YES votes among members according to the VotingPolicy specified for the Network. The specified ProposalActions are not carried out.

EXPIRED - Members did not cast the number of votes required to determine the proposal outcome before the proposal expired. The specified ProposalActions are not carried out.

ACTION_FAILED - One or more of the specified ProposalActions in a proposal that was approved could not be completed because of an error.

proposalSummary_arn :: Lens' ProposalSummary (Maybe Text)

The Amazon Resource Name (ARN) of the proposal. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

proposalSummary_proposedByMemberId :: Lens' ProposalSummary (Maybe Text)

The unique identifier of the member that created the proposal.

proposalSummary_proposalId :: Lens' ProposalSummary (Maybe Text)

The unique identifier of the proposal.

proposalSummary_proposedByMemberName :: Lens' ProposalSummary (Maybe Text)

The name of the member that created the proposal.

proposalSummary_expirationDate :: Lens' ProposalSummary (Maybe UTCTime)

The date and time that the proposal expires. This is the CreationDate plus the ProposalDurationInHours that is specified in the ProposalThresholdPolicy. After this date and time, if members have not cast enough votes to determine the outcome according to the voting policy, the proposal is EXPIRED and Actions are not carried out.

proposalSummary_creationDate :: Lens' ProposalSummary (Maybe UTCTime)

The date and time that the proposal was created.

proposalSummary_description :: Lens' ProposalSummary (Maybe Text)

The description of the proposal.

RemoveAction

removeAction_memberId :: Lens' RemoveAction Text

The unique identifier of the member to remove.

VoteSummary

voteSummary_memberName :: Lens' VoteSummary (Maybe Text)

The name of the member that cast the vote.

voteSummary_memberId :: Lens' VoteSummary (Maybe Text)

The unique identifier of the member that cast the vote.

voteSummary_vote :: Lens' VoteSummary (Maybe VoteValue)

The vote value, either YES or NO.

VotingPolicy

votingPolicy_approvalThresholdPolicy :: Lens' VotingPolicy (Maybe ApprovalThresholdPolicy)

Defines the rules for the network for voting on proposals, such as the percentage of YES votes required for the proposal to be approved and the duration of the proposal. The policy applies to all proposals and is specified when the network is created.