Network.AWS.ManagedBlockChain.RejectInvitation

Description

Rejects an invitation to join a network. This action can be called by a principal in an AWS account that has received an invitation to create a member and join a network.

Applies only to Hyperledger Fabric.

Creating a Request

data RejectInvitation

See: newRejectInvitation smart constructor.

Constructors

RejectInvitation'

Fields

invitationId :: Text

The unique identifier of the invitation to reject.

newRejectInvitation

Arguments

:: Text
$sel:invitationId:RejectInvitation'

-> RejectInvitation

Create a value of RejectInvitation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:invitationId:RejectInvitation', rejectInvitation_invitationId - The unique identifier of the invitation to reject.

Request Lenses

rejectInvitation_invitationId :: Lens' RejectInvitation Text

The unique identifier of the invitation to reject.

Destructuring the Response

data RejectInvitationResponse

See: newRejectInvitationResponse smart constructor.

Constructors

RejectInvitationResponse'

Fields

httpStatus :: Int

The response's http status code.

newRejectInvitationResponse

Arguments

:: Int
$sel:httpStatus:RejectInvitationResponse'

-> RejectInvitationResponse

Create a value of RejectInvitationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:RejectInvitationResponse', rejectInvitationResponse_httpStatus - The response's http status code.

Response Lenses

rejectInvitationResponse_httpStatus :: Lens' RejectInvitationResponse Int

The response's http status code.