Network.AWS.ManagedBlockChain.Types.MemberConfiguration

Description

Documentation

data MemberConfiguration

Configuration properties of the member.

Applies only to Hyperledger Fabric.

See: newMemberConfiguration smart constructor.

Constructors

MemberConfiguration'

Fields

kmsKeyArn :: Maybe Text

The Amazon Resource Name (ARN) of the customer managed key in AWS Key Management Service (AWS KMS) to use for encryption at rest in the member. This parameter is inherited by any nodes that this member creates.

Use one of the following options to specify this parameter:

Undefined or empty string - The member uses an AWS owned KMS key for encryption by default.

A valid symmetric customer managed KMS key - The member uses the specified key for encryption.

Amazon Managed Blockchain doesn't support asymmetric keys. For more information, see Using symmetric and asymmetric keys in the AWS Key Management Service Developer Guide.

The following is an example of a KMS key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

logPublishingConfiguration :: Maybe MemberLogPublishingConfiguration

Configuration properties for logging events associated with a member of a Managed Blockchain network.

description :: Maybe Text

An optional description of the member.

tags :: Maybe (HashMap Text Text)

Tags assigned to the member. Tags consist of a key and optional value. For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

When specifying tags during creation, you can specify multiple key-value pairs in a single request, with an overall maximum of 50 tags added to each resource.

name :: Text

The name of the member.

frameworkConfiguration :: MemberFrameworkConfiguration

Configuration properties of the blockchain framework relevant to the member.

newMemberConfiguration

Arguments

:: Text
$sel:name:MemberConfiguration'

-> MemberFrameworkConfiguration
$sel:frameworkConfiguration:MemberConfiguration'

-> MemberConfiguration

Create a value of MemberConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyArn:MemberConfiguration', memberConfiguration_kmsKeyArn - The Amazon Resource Name (ARN) of the customer managed key in AWS Key Management Service (AWS KMS) to use for encryption at rest in the member. This parameter is inherited by any nodes that this member creates.

Use one of the following options to specify this parameter:

Undefined or empty string - The member uses an AWS owned KMS key for encryption by default.

A valid symmetric customer managed KMS key - The member uses the specified key for encryption.

Amazon Managed Blockchain doesn't support asymmetric keys. For more information, see Using symmetric and asymmetric keys in the AWS Key Management Service Developer Guide.

The following is an example of a KMS key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

$sel:logPublishingConfiguration:MemberConfiguration', memberConfiguration_logPublishingConfiguration - Configuration properties for logging events associated with a member of a Managed Blockchain network.

$sel:description:MemberConfiguration', memberConfiguration_description - An optional description of the member.

$sel:tags:MemberConfiguration', memberConfiguration_tags - Tags assigned to the member. Tags consist of a key and optional value. For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

When specifying tags during creation, you can specify multiple key-value pairs in a single request, with an overall maximum of 50 tags added to each resource.

$sel:name:MemberConfiguration', memberConfiguration_name - The name of the member.

$sel:frameworkConfiguration:MemberConfiguration', memberConfiguration_frameworkConfiguration - Configuration properties of the blockchain framework relevant to the member.

memberConfiguration_kmsKeyArn :: Lens' MemberConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the customer managed key in AWS Key Management Service (AWS KMS) to use for encryption at rest in the member. This parameter is inherited by any nodes that this member creates.

Use one of the following options to specify this parameter:

Undefined or empty string - The member uses an AWS owned KMS key for encryption by default.

A valid symmetric customer managed KMS key - The member uses the specified key for encryption.

Amazon Managed Blockchain doesn't support asymmetric keys. For more information, see Using symmetric and asymmetric keys in the AWS Key Management Service Developer Guide.

The following is an example of a KMS key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

memberConfiguration_logPublishingConfiguration :: Lens' MemberConfiguration (Maybe MemberLogPublishingConfiguration)

Configuration properties for logging events associated with a member of a Managed Blockchain network.

memberConfiguration_description :: Lens' MemberConfiguration (Maybe Text)

An optional description of the member.

memberConfiguration_tags :: Lens' MemberConfiguration (Maybe (HashMap Text Text))

Tags assigned to the member. Tags consist of a key and optional value. For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

When specifying tags during creation, you can specify multiple key-value pairs in a single request, with an overall maximum of 50 tags added to each resource.

memberConfiguration_name :: Lens' MemberConfiguration Text

The name of the member.

memberConfiguration_frameworkConfiguration :: Lens' MemberConfiguration MemberFrameworkConfiguration

Configuration properties of the blockchain framework relevant to the member.