Network.AWS.ManagedBlockChain.Types.Network

Description

Documentation

data Network

Network configuration properties.

See: newNetwork smart constructor.

Constructors

Network'

Fields

status :: Maybe NetworkStatus

The current status of the network.

framework :: Maybe Framework

The blockchain framework that the network uses.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the network. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

frameworkVersion :: Maybe Text

The version of the blockchain framework that the network uses.

vpcEndpointServiceName :: Maybe Text

The VPC endpoint service name of the VPC endpoint service of the network. Members use the VPC endpoint service name to create a VPC endpoint to access network resources.

name :: Maybe Text

The name of the network.

id :: Maybe Text

The unique identifier of the network.

votingPolicy :: Maybe VotingPolicy

The voting rules for the network to decide if a proposal is accepted.

creationDate :: Maybe POSIX

The date and time that the network was created.

frameworkAttributes :: Maybe NetworkFrameworkAttributes

Attributes of the blockchain framework that the network uses.

description :: Maybe Text

Attributes of the blockchain framework for the network.

tags :: Maybe (HashMap Text Text)

Tags assigned to the network. Each tag consists of a key and optional value.

For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Ethereum Developer Guide, or Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

newNetwork :: Network

Create a value of Network with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Network', network_status - The current status of the network.

$sel:framework:Network', network_framework - The blockchain framework that the network uses.

$sel:arn:Network', network_arn - The Amazon Resource Name (ARN) of the network. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

$sel:frameworkVersion:Network', network_frameworkVersion - The version of the blockchain framework that the network uses.

$sel:vpcEndpointServiceName:Network', network_vpcEndpointServiceName - The VPC endpoint service name of the VPC endpoint service of the network. Members use the VPC endpoint service name to create a VPC endpoint to access network resources.

$sel:name:Network', network_name - The name of the network.

$sel:id:Network', network_id - The unique identifier of the network.

$sel:votingPolicy:Network', network_votingPolicy - The voting rules for the network to decide if a proposal is accepted.

$sel:creationDate:Network', network_creationDate - The date and time that the network was created.

$sel:frameworkAttributes:Network', network_frameworkAttributes - Attributes of the blockchain framework that the network uses.

$sel:description:Network', network_description - Attributes of the blockchain framework for the network.

$sel:tags:Network', network_tags - Tags assigned to the network. Each tag consists of a key and optional value.

For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Ethereum Developer Guide, or Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

network_status :: Lens' Network (Maybe NetworkStatus)

The current status of the network.

network_framework :: Lens' Network (Maybe Framework)

The blockchain framework that the network uses.

network_arn :: Lens' Network (Maybe Text)

The Amazon Resource Name (ARN) of the network. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

network_frameworkVersion :: Lens' Network (Maybe Text)

The version of the blockchain framework that the network uses.

network_vpcEndpointServiceName :: Lens' Network (Maybe Text)

The VPC endpoint service name of the VPC endpoint service of the network. Members use the VPC endpoint service name to create a VPC endpoint to access network resources.

network_name :: Lens' Network (Maybe Text)

The name of the network.

network_id :: Lens' Network (Maybe Text)

The unique identifier of the network.

network_votingPolicy :: Lens' Network (Maybe VotingPolicy)

The voting rules for the network to decide if a proposal is accepted.

network_creationDate :: Lens' Network (Maybe UTCTime)

The date and time that the network was created.

network_frameworkAttributes :: Lens' Network (Maybe NetworkFrameworkAttributes)

Attributes of the blockchain framework that the network uses.

network_description :: Lens' Network (Maybe Text)

Attributes of the blockchain framework for the network.

network_tags :: Lens' Network (Maybe (HashMap Text Text))

Tags assigned to the network. Each tag consists of a key and optional value.

For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Ethereum Developer Guide, or Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.