Network.AWS.ManagedBlockChain.Types.Node

Description

Documentation

data Node

Configuration properties of a node.

See: newNode smart constructor.

Constructors

Node'

Fields

status :: Maybe NodeStatus

The status of the node.

CREATING - The AWS account is in the process of creating a node.

AVAILABLE - The node has been created and can participate in the network.

UNHEALTHY - The node is impaired and might not function as expected. Amazon Managed Blockchain automatically finds nodes in this state and tries to recover them. If a node is recoverable, it returns to AVAILABLE. Otherwise, it moves to FAILED status.

CREATE_FAILED - The AWS account attempted to create a node and creation failed.

UPDATING - The node is in the process of being updated.

DELETING - The node is in the process of being deleted.

DELETED - The node can no longer participate on the network.

FAILED - The node is no longer functional, cannot be recovered, and must be deleted.

INACCESSIBLE_ENCRYPTION_KEY - The node is impaired and might not function as expected because it cannot access the specified customer managed key in AWS KMS for encryption at rest. Either the KMS key was disabled or deleted, or the grants on the key were revoked.

The effect of disabling or deleting a key, or revoking a grant is not immediate. The node resource might take some time to find that the key is inaccessible. When a resource is in this state, we recommend deleting and recreating the resource.

kmsKeyArn :: Maybe Text

The Amazon Resource Name (ARN) of the customer managed key in AWS Key Management Service (AWS KMS) that the node uses for encryption at rest. If the value of this parameter is "AWS Owned KMS Key", the node uses an AWS owned KMS key for encryption. The node inherits this parameter from the member that it belongs to.

Applies only to Hyperledger Fabric.

logPublishingConfiguration :: Maybe NodeLogPublishingConfiguration

Configuration properties for logging events associated with a peer node on a Hyperledger Fabric network on Managed Blockchain.

memberId :: Maybe Text

The unique identifier of the member to which the node belongs.

Applies only to Hyperledger Fabric.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the node. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

networkId :: Maybe Text

The unique identifier of the network that the node is on.

instanceType :: Maybe Text

The instance type of the node.

stateDB :: Maybe StateDBType

The state database that the node uses. Values are LevelDB or CouchDB.

Applies only to Hyperledger Fabric.

availabilityZone :: Maybe Text

The Availability Zone in which the node exists. Required for Ethereum nodes.

id :: Maybe Text

The unique identifier of the node.

creationDate :: Maybe POSIX

The date and time that the node was created.

frameworkAttributes :: Maybe NodeFrameworkAttributes

Attributes of the blockchain framework being used.

tags :: Maybe (HashMap Text Text)

Tags assigned to the node. Each tag consists of a key and optional value.

For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Ethereum Developer Guide, or Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

newNode :: Node

Create a value of Node with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Node', node_status - The status of the node.

CREATING - The AWS account is in the process of creating a node.

AVAILABLE - The node has been created and can participate in the network.

UNHEALTHY - The node is impaired and might not function as expected. Amazon Managed Blockchain automatically finds nodes in this state and tries to recover them. If a node is recoverable, it returns to AVAILABLE. Otherwise, it moves to FAILED status.

CREATE_FAILED - The AWS account attempted to create a node and creation failed.

UPDATING - The node is in the process of being updated.

DELETING - The node is in the process of being deleted.

DELETED - The node can no longer participate on the network.

FAILED - The node is no longer functional, cannot be recovered, and must be deleted.

INACCESSIBLE_ENCRYPTION_KEY - The node is impaired and might not function as expected because it cannot access the specified customer managed key in AWS KMS for encryption at rest. Either the KMS key was disabled or deleted, or the grants on the key were revoked.

The effect of disabling or deleting a key, or revoking a grant is not immediate. The node resource might take some time to find that the key is inaccessible. When a resource is in this state, we recommend deleting and recreating the resource.

$sel:kmsKeyArn:Node', node_kmsKeyArn - The Amazon Resource Name (ARN) of the customer managed key in AWS Key Management Service (AWS KMS) that the node uses for encryption at rest. If the value of this parameter is "AWS Owned KMS Key", the node uses an AWS owned KMS key for encryption. The node inherits this parameter from the member that it belongs to.

Applies only to Hyperledger Fabric.

$sel:logPublishingConfiguration:Node', node_logPublishingConfiguration - Configuration properties for logging events associated with a peer node on a Hyperledger Fabric network on Managed Blockchain.

$sel:memberId:Node', node_memberId - The unique identifier of the member to which the node belongs.

Applies only to Hyperledger Fabric.

$sel:arn:Node', node_arn - The Amazon Resource Name (ARN) of the node. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

$sel:networkId:Node', node_networkId - The unique identifier of the network that the node is on.

$sel:instanceType:Node', node_instanceType - The instance type of the node.

$sel:stateDB:Node', node_stateDB - The state database that the node uses. Values are LevelDB or CouchDB.

Applies only to Hyperledger Fabric.

$sel:availabilityZone:Node', node_availabilityZone - The Availability Zone in which the node exists. Required for Ethereum nodes.

$sel:id:Node', node_id - The unique identifier of the node.

$sel:creationDate:Node', node_creationDate - The date and time that the node was created.

$sel:frameworkAttributes:Node', node_frameworkAttributes - Attributes of the blockchain framework being used.

$sel:tags:Node', node_tags - Tags assigned to the node. Each tag consists of a key and optional value.

For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Ethereum Developer Guide, or Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

node_status :: Lens' Node (Maybe NodeStatus)

The status of the node.

CREATING - The AWS account is in the process of creating a node.

AVAILABLE - The node has been created and can participate in the network.

UNHEALTHY - The node is impaired and might not function as expected. Amazon Managed Blockchain automatically finds nodes in this state and tries to recover them. If a node is recoverable, it returns to AVAILABLE. Otherwise, it moves to FAILED status.

CREATE_FAILED - The AWS account attempted to create a node and creation failed.

UPDATING - The node is in the process of being updated.

DELETING - The node is in the process of being deleted.

DELETED - The node can no longer participate on the network.

FAILED - The node is no longer functional, cannot be recovered, and must be deleted.

INACCESSIBLE_ENCRYPTION_KEY - The node is impaired and might not function as expected because it cannot access the specified customer managed key in AWS KMS for encryption at rest. Either the KMS key was disabled or deleted, or the grants on the key were revoked.

The effect of disabling or deleting a key, or revoking a grant is not immediate. The node resource might take some time to find that the key is inaccessible. When a resource is in this state, we recommend deleting and recreating the resource.

node_kmsKeyArn :: Lens' Node (Maybe Text)

The Amazon Resource Name (ARN) of the customer managed key in AWS Key Management Service (AWS KMS) that the node uses for encryption at rest. If the value of this parameter is "AWS Owned KMS Key", the node uses an AWS owned KMS key for encryption. The node inherits this parameter from the member that it belongs to.

Applies only to Hyperledger Fabric.

node_logPublishingConfiguration :: Lens' Node (Maybe NodeLogPublishingConfiguration)

Configuration properties for logging events associated with a peer node on a Hyperledger Fabric network on Managed Blockchain.

node_memberId :: Lens' Node (Maybe Text)

The unique identifier of the member to which the node belongs.

Applies only to Hyperledger Fabric.

node_arn :: Lens' Node (Maybe Text)

The Amazon Resource Name (ARN) of the node. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the AWS General Reference.

node_networkId :: Lens' Node (Maybe Text)

The unique identifier of the network that the node is on.

node_instanceType :: Lens' Node (Maybe Text)

The instance type of the node.

node_stateDB :: Lens' Node (Maybe StateDBType)

The state database that the node uses. Values are LevelDB or CouchDB.

Applies only to Hyperledger Fabric.

node_availabilityZone :: Lens' Node (Maybe Text)

The Availability Zone in which the node exists. Required for Ethereum nodes.

node_id :: Lens' Node (Maybe Text)

The unique identifier of the node.

node_creationDate :: Lens' Node (Maybe UTCTime)

The date and time that the node was created.

node_frameworkAttributes :: Lens' Node (Maybe NodeFrameworkAttributes)

Attributes of the blockchain framework being used.

node_tags :: Lens' Node (Maybe (HashMap Text Text))

Tags assigned to the node. Each tag consists of a key and optional value.

For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Ethereum Developer Guide, or Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.