Network.AWS.ManagedBlockChain.VoteOnProposal

Description

Casts a vote for a specified ProposalId on behalf of a member. The member to vote as, specified by VoterMemberId, must be in the same AWS account as the principal that calls the action.

Applies only to Hyperledger Fabric.

Creating a Request

data VoteOnProposal

See: newVoteOnProposal smart constructor.

Constructors

VoteOnProposal'

Fields

networkId :: Text

The unique identifier of the network.

proposalId :: Text

The unique identifier of the proposal.

voterMemberId :: Text

The unique identifier of the member casting the vote.

vote :: VoteValue

The value of the vote.

newVoteOnProposal

Arguments

:: Text
$sel:networkId:VoteOnProposal'

-> Text
$sel:proposalId:VoteOnProposal'

-> Text
$sel:voterMemberId:VoteOnProposal'

-> VoteValue
$sel:vote:VoteOnProposal'

-> VoteOnProposal

Create a value of VoteOnProposal with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkId:VoteOnProposal', voteOnProposal_networkId - The unique identifier of the network.

$sel:proposalId:VoteOnProposal', voteOnProposal_proposalId - The unique identifier of the proposal.

$sel:voterMemberId:VoteOnProposal', voteOnProposal_voterMemberId - The unique identifier of the member casting the vote.

$sel:vote:VoteOnProposal', voteOnProposal_vote - The value of the vote.

Request Lenses

voteOnProposal_networkId :: Lens' VoteOnProposal Text

The unique identifier of the network.

voteOnProposal_proposalId :: Lens' VoteOnProposal Text

The unique identifier of the proposal.

voteOnProposal_voterMemberId :: Lens' VoteOnProposal Text

The unique identifier of the member casting the vote.

voteOnProposal_vote :: Lens' VoteOnProposal VoteValue

The value of the vote.

Destructuring the Response

data VoteOnProposalResponse

See: newVoteOnProposalResponse smart constructor.

Constructors

VoteOnProposalResponse'

Fields

httpStatus :: Int

The response's http status code.

newVoteOnProposalResponse

Arguments

:: Int
$sel:httpStatus:VoteOnProposalResponse'

-> VoteOnProposalResponse

Create a value of VoteOnProposalResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:VoteOnProposalResponse', voteOnProposalResponse_httpStatus - The response's http status code.

Response Lenses

voteOnProposalResponse_httpStatus :: Lens' VoteOnProposalResponse Int

The response's http status code.