Network.AWS.MarketplaceAnalytics

Description

Derived from API version 2015-07-01 of the AWS service descriptions, licensed under Apache 2.0.

Provides AWS Marketplace business intelligence data on-demand.

Service Configuration

defaultService :: Service

API version 2015-07-01 of the Amazon Marketplace Commerce Analytics SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by MarketplaceAnalytics.

MarketplaceCommerceAnalyticsException

_MarketplaceCommerceAnalyticsException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown when an internal service error occurs.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

StartSupportDataExport

data StartSupportDataExport

Container for the parameters to the StartSupportDataExport operation.

See: newStartSupportDataExport smart constructor.

Constructors

StartSupportDataExport' (Maybe (HashMap Text Text)) (Maybe Text) SupportDataSetType POSIX Text Text Text

newStartSupportDataExport

Arguments

:: SupportDataSetType
$sel:dataSetType:StartSupportDataExport'

-> UTCTime
$sel:fromDate:StartSupportDataExport'

-> Text
$sel:roleNameArn:StartSupportDataExport'

-> Text
$sel:destinationS3BucketName:StartSupportDataExport'

-> Text
$sel:snsTopicArn:StartSupportDataExport'

-> StartSupportDataExport

Create a value of StartSupportDataExport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customerDefinedValues:StartSupportDataExport', startSupportDataExport_customerDefinedValues - (Optional) Key-value pairs which will be returned, unmodified, in the Amazon SNS notification message and the data set metadata file.

$sel:destinationS3Prefix:StartSupportDataExport', startSupportDataExport_destinationS3Prefix - (Optional) The desired S3 prefix for the published data set, similar to a directory path in standard file systems. For example, if given the bucket name "mybucket" and the prefix "myprefix/mydatasets", the output file "outputfile" would be published to "s3://mybucket/myprefix/mydatasets/outputfile". If the prefix directory structure does not exist, it will be created. If no prefix is provided, the data set will be published to the S3 bucket root.

$sel:dataSetType:StartSupportDataExport', startSupportDataExport_dataSetType - Specifies the data set type to be written to the output csv file. The data set types customer_support_contacts_data and test_customer_support_contacts_data both result in a csv file containing the following fields: Product Id, Product Code, Customer Guid, Subscription Guid, Subscription Start Date, Organization, AWS Account Id, Given Name, Surname, Telephone Number, Email, Title, Country Code, ZIP Code, Operation Type, and Operation Time.

customer_support_contacts_data Customer support contact data. The data set will contain all changes (Creates, Updates, and Deletes) to customer support contact data from the date specified in the from_date parameter.

test_customer_support_contacts_data An example data set containing static test data in the same format as customer_support_contacts_data

$sel:fromDate:StartSupportDataExport', startSupportDataExport_fromDate - The start date from which to retrieve the data set in UTC. This parameter only affects the customer_support_contacts_data data set type.

$sel:roleNameArn:StartSupportDataExport', startSupportDataExport_roleNameArn - The Amazon Resource Name (ARN) of the Role with an attached permissions policy to interact with the provided AWS services.

$sel:destinationS3BucketName:StartSupportDataExport', startSupportDataExport_destinationS3BucketName - The name (friendly name, not ARN) of the destination S3 bucket.

$sel:snsTopicArn:StartSupportDataExport', startSupportDataExport_snsTopicArn - Amazon Resource Name (ARN) for the SNS Topic that will be notified when the data set has been published or if an error has occurred.

data StartSupportDataExportResponse

Container for the result of the StartSupportDataExport operation.

See: newStartSupportDataExportResponse smart constructor.

Constructors

StartSupportDataExportResponse' (Maybe Text) Int

newStartSupportDataExportResponse

Arguments

:: Int
$sel:httpStatus:StartSupportDataExportResponse'

-> StartSupportDataExportResponse

Create a value of StartSupportDataExportResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataSetRequestId:StartSupportDataExportResponse', startSupportDataExportResponse_dataSetRequestId - A unique identifier representing a specific request to the StartSupportDataExport operation. This identifier can be used to correlate a request with notifications from the SNS topic.

$sel:httpStatus:StartSupportDataExportResponse', startSupportDataExportResponse_httpStatus - The response's http status code.

GenerateDataSet

data GenerateDataSet

Container for the parameters to the GenerateDataSet operation.

See: newGenerateDataSet smart constructor.

Constructors

GenerateDataSet' (Maybe (HashMap Text Text)) (Maybe Text) DataSetType POSIX Text Text Text

newGenerateDataSet

Arguments

:: DataSetType
$sel:dataSetType:GenerateDataSet'

-> UTCTime
$sel:dataSetPublicationDate:GenerateDataSet'

-> Text
$sel:roleNameArn:GenerateDataSet'

-> Text
$sel:destinationS3BucketName:GenerateDataSet'

-> Text
$sel:snsTopicArn:GenerateDataSet'

-> GenerateDataSet

Create a value of GenerateDataSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customerDefinedValues:GenerateDataSet', generateDataSet_customerDefinedValues - (Optional) Key-value pairs which will be returned, unmodified, in the Amazon SNS notification message and the data set metadata file. These key-value pairs can be used to correlated responses with tracking information from other systems.

$sel:destinationS3Prefix:GenerateDataSet', generateDataSet_destinationS3Prefix - (Optional) The desired S3 prefix for the published data set, similar to a directory path in standard file systems. For example, if given the bucket name "mybucket" and the prefix "myprefix/mydatasets", the output file "outputfile" would be published to "s3://mybucket/myprefix/mydatasets/outputfile". If the prefix directory structure does not exist, it will be created. If no prefix is provided, the data set will be published to the S3 bucket root.

$sel:dataSetType:GenerateDataSet', generateDataSet_dataSetType - The desired data set type.

customer_subscriber_hourly_monthly_subscriptions

From 2017-09-15 to present: Available daily by 24:00 UTC.

customer_subscriber_annual_subscriptions

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_usage_by_instance_type

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_fees

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_free_trial_conversions

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_new_instances

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_new_product_subscribers

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_canceled_product_subscribers

From 2017-09-15 to present: Available daily by 24:00 UTC.

monthly_revenue_billing_and_revenue_data

From 2017-09-15 to present: Available monthly on the 15th day of the month by 24:00 UTC. Data includes metered transactions (e.g. hourly) from one month prior.

monthly_revenue_annual_subscriptions

From 2017-09-15 to present: Available monthly on the 15th day of the month by 24:00 UTC. Data includes up-front software charges (e.g. annual) from one month prior.

monthly_revenue_field_demonstration_usage

From 2018-03-15 to present: Available monthly on the 15th day of the month by 24:00 UTC.

monthly_revenue_flexible_payment_schedule

From 2018-11-15 to present: Available monthly on the 15th day of the month by 24:00 UTC.

disbursed_amount_by_product

From 2017-09-15 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_instance_hours

From 2017-09-15 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_customer_geo

From 2017-09-15 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_age_of_uncollected_funds

From 2017-09-15 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_age_of_disbursed_funds

From 2017-09-15 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_age_of_past_due_funds

From 2018-04-07 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_uncollected_funds_breakdown

From 2019-10-04 to present: Available every 30 days by 24:00 UTC.

sales_compensation_billed_revenue

From 2017-09-15 to present: Available monthly on the 15th day of the month by 24:00 UTC. Data includes metered transactions (e.g. hourly) from one month prior, and up-front software charges (e.g. annual) from one month prior.

us_sales_and_use_tax_records

From 2017-09-15 to present: Available monthly on the 15th day of the month by 24:00 UTC.

disbursed_amount_by_product_with_uncollected_funds

This data set is deprecated. Download related reports from AMMP instead!

customer_profile_by_industry

This data set is deprecated. Download related reports from AMMP instead!

customer_profile_by_revenue

This data set is deprecated. Download related reports from AMMP instead!

customer_profile_by_geography

This data set is deprecated. Download related reports from AMMP instead!

$sel:dataSetPublicationDate:GenerateDataSet', generateDataSet_dataSetPublicationDate - The date a data set was published. For daily data sets, provide a date with day-level granularity for the desired day. For monthly data sets except those with prefix disbursed_amount, provide a date with month-level granularity for the desired month (the day value will be ignored). For data sets with prefix disbursed_amount, provide a date with day-level granularity for the desired day. For these data sets we will look backwards in time over the range of 31 days until the first data set is found (the latest one).

$sel:roleNameArn:GenerateDataSet', generateDataSet_roleNameArn - The Amazon Resource Name (ARN) of the Role with an attached permissions policy to interact with the provided AWS services.

$sel:destinationS3BucketName:GenerateDataSet', generateDataSet_destinationS3BucketName - The name (friendly name, not ARN) of the destination S3 bucket.

$sel:snsTopicArn:GenerateDataSet', generateDataSet_snsTopicArn - Amazon Resource Name (ARN) for the SNS Topic that will be notified when the data set has been published or if an error has occurred.

data GenerateDataSetResponse

Container for the result of the GenerateDataSet operation.

See: newGenerateDataSetResponse smart constructor.

Constructors

GenerateDataSetResponse' (Maybe Text) Int

newGenerateDataSetResponse

Arguments

:: Int
$sel:httpStatus:GenerateDataSetResponse'

-> GenerateDataSetResponse

Create a value of GenerateDataSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataSetRequestId:GenerateDataSetResponse', generateDataSetResponse_dataSetRequestId - A unique identifier representing a specific request to the GenerateDataSet operation. This identifier can be used to correlate a request with notifications from the SNS topic.

$sel:httpStatus:GenerateDataSetResponse', generateDataSetResponse_httpStatus - The response's http status code.

Types

DataSetType

newtype DataSetType

Constructors

DataSetType'

Fields

fromDataSetType :: Text

Bundled Patterns

pattern DataSetType_Customer_profile_by_geography :: DataSetType

pattern DataSetType_Customer_profile_by_industry :: DataSetType

pattern DataSetType_Customer_profile_by_revenue :: DataSetType

pattern DataSetType_Customer_subscriber_annual_subscriptions :: DataSetType

pattern DataSetType_Customer_subscriber_hourly_monthly_subscriptions :: DataSetType

pattern DataSetType_Daily_business_canceled_product_subscribers :: DataSetType

pattern DataSetType_Daily_business_fees :: DataSetType

pattern DataSetType_Daily_business_free_trial_conversions :: DataSetType

pattern DataSetType_Daily_business_new_instances :: DataSetType

pattern DataSetType_Daily_business_new_product_subscribers :: DataSetType

pattern DataSetType_Daily_business_usage_by_instance_type :: DataSetType

pattern DataSetType_Disbursed_amount_by_age_of_disbursed_funds :: DataSetType

pattern DataSetType_Disbursed_amount_by_age_of_past_due_funds :: DataSetType

pattern DataSetType_Disbursed_amount_by_age_of_uncollected_funds :: DataSetType

pattern DataSetType_Disbursed_amount_by_customer_geo :: DataSetType

pattern DataSetType_Disbursed_amount_by_instance_hours :: DataSetType

pattern DataSetType_Disbursed_amount_by_product :: DataSetType

pattern DataSetType_Disbursed_amount_by_product_with_uncollected_funds :: DataSetType

pattern DataSetType_Disbursed_amount_by_uncollected_funds_breakdown :: DataSetType

pattern DataSetType_Monthly_revenue_annual_subscriptions :: DataSetType

pattern DataSetType_Monthly_revenue_billing_and_revenue_data :: DataSetType

pattern DataSetType_Monthly_revenue_field_demonstration_usage :: DataSetType

pattern DataSetType_Monthly_revenue_flexible_payment_schedule :: DataSetType

pattern DataSetType_Sales_compensation_billed_revenue :: DataSetType

pattern DataSetType_Us_sales_and_use_tax_records :: DataSetType

SupportDataSetType

newtype SupportDataSetType

Constructors

SupportDataSetType'

Fields

fromSupportDataSetType :: Text

Bundled Patterns

pattern SupportDataSetType_Customer_support_contacts_data :: SupportDataSetType

pattern SupportDataSetType_Test_customer_support_contacts_data :: SupportDataSetType