Network.AWS.MarketplaceAnalytics.GenerateDataSet

Description

Given a data set type and data set publication date, asynchronously publishes the requested data set to the specified S3 bucket and notifies the specified SNS topic once the data is available. Returns a unique request identifier that can be used to correlate requests with notifications from the SNS topic. Data sets will be published in comma-separated values (CSV) format with the file name {data_set_type}_YYYY-MM-DD.csv. If a file with the same name already exists (e.g. if the same data set is requested twice), the original file will be overwritten by the new file. Requires a Role with an attached permissions policy providing Allow permissions for the following actions: s3:PutObject, s3:GetBucketLocation, sns:GetTopicAttributes, sns:Publish, iam:GetRolePolicy.

Creating a Request

data GenerateDataSet

Container for the parameters to the GenerateDataSet operation.

See: newGenerateDataSet smart constructor.

Constructors

GenerateDataSet'

Fields

destinationS3Prefix :: Maybe Text

(Optional) The desired S3 prefix for the published data set, similar to a directory path in standard file systems. For example, if given the bucket name "mybucket" and the prefix "myprefix/mydatasets", the output file "outputfile" would be published to "s3://mybucket/myprefix/mydatasets/outputfile". If the prefix directory structure does not exist, it will be created. If no prefix is provided, the data set will be published to the S3 bucket root.

customerDefinedValues :: Maybe (HashMap Text Text)

(Optional) Key-value pairs which will be returned, unmodified, in the Amazon SNS notification message and the data set metadata file. These key-value pairs can be used to correlated responses with tracking information from other systems.

dataSetType :: DataSetType

The desired data set type.

customer_subscriber_hourly_monthly_subscriptions

From 2017-09-15 to present: Available daily by 24:00 UTC.

customer_subscriber_annual_subscriptions

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_usage_by_instance_type

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_fees

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_free_trial_conversions

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_new_instances

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_new_product_subscribers

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_canceled_product_subscribers

From 2017-09-15 to present: Available daily by 24:00 UTC.

monthly_revenue_billing_and_revenue_data

From 2017-09-15 to present: Available monthly on the 15th day of the month by 24:00 UTC. Data includes metered transactions (e.g. hourly) from one month prior.

monthly_revenue_annual_subscriptions

From 2017-09-15 to present: Available monthly on the 15th day of the month by 24:00 UTC. Data includes up-front software charges (e.g. annual) from one month prior.

monthly_revenue_field_demonstration_usage

From 2018-03-15 to present: Available monthly on the 15th day of the month by 24:00 UTC.

monthly_revenue_flexible_payment_schedule

From 2018-11-15 to present: Available monthly on the 15th day of the month by 24:00 UTC.

disbursed_amount_by_product

From 2017-09-15 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_instance_hours

From 2017-09-15 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_customer_geo

From 2017-09-15 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_age_of_uncollected_funds

From 2017-09-15 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_age_of_disbursed_funds

From 2017-09-15 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_age_of_past_due_funds

From 2018-04-07 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_uncollected_funds_breakdown

From 2019-10-04 to present: Available every 30 days by 24:00 UTC.

sales_compensation_billed_revenue

From 2017-09-15 to present: Available monthly on the 15th day of the month by 24:00 UTC. Data includes metered transactions (e.g. hourly) from one month prior, and up-front software charges (e.g. annual) from one month prior.

us_sales_and_use_tax_records

From 2017-09-15 to present: Available monthly on the 15th day of the month by 24:00 UTC.

disbursed_amount_by_product_with_uncollected_funds

This data set is deprecated. Download related reports from AMMP instead!

customer_profile_by_industry

This data set is deprecated. Download related reports from AMMP instead!

customer_profile_by_revenue

This data set is deprecated. Download related reports from AMMP instead!

customer_profile_by_geography

This data set is deprecated. Download related reports from AMMP instead!

dataSetPublicationDate :: POSIX

The date a data set was published. For daily data sets, provide a date with day-level granularity for the desired day. For monthly data sets except those with prefix disbursed_amount, provide a date with month-level granularity for the desired month (the day value will be ignored). For data sets with prefix disbursed_amount, provide a date with day-level granularity for the desired day. For these data sets we will look backwards in time over the range of 31 days until the first data set is found (the latest one).

roleNameArn :: Text

The Amazon Resource Name (ARN) of the Role with an attached permissions policy to interact with the provided AWS services.

destinationS3BucketName :: Text

The name (friendly name, not ARN) of the destination S3 bucket.

snsTopicArn :: Text

Amazon Resource Name (ARN) for the SNS Topic that will be notified when the data set has been published or if an error has occurred.

newGenerateDataSet

Arguments

:: DataSetType
$sel:dataSetType:GenerateDataSet'

-> UTCTime
$sel:dataSetPublicationDate:GenerateDataSet'

-> Text
$sel:roleNameArn:GenerateDataSet'

-> Text
$sel:destinationS3BucketName:GenerateDataSet'

-> Text
$sel:snsTopicArn:GenerateDataSet'

-> GenerateDataSet

Create a value of GenerateDataSet with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationS3Prefix:GenerateDataSet', generateDataSet_destinationS3Prefix - (Optional) The desired S3 prefix for the published data set, similar to a directory path in standard file systems. For example, if given the bucket name "mybucket" and the prefix "myprefix/mydatasets", the output file "outputfile" would be published to "s3://mybucket/myprefix/mydatasets/outputfile". If the prefix directory structure does not exist, it will be created. If no prefix is provided, the data set will be published to the S3 bucket root.

$sel:customerDefinedValues:GenerateDataSet', generateDataSet_customerDefinedValues - (Optional) Key-value pairs which will be returned, unmodified, in the Amazon SNS notification message and the data set metadata file. These key-value pairs can be used to correlated responses with tracking information from other systems.

$sel:dataSetType:GenerateDataSet', generateDataSet_dataSetType - The desired data set type.

customer_subscriber_hourly_monthly_subscriptions

From 2017-09-15 to present: Available daily by 24:00 UTC.

customer_subscriber_annual_subscriptions

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_usage_by_instance_type

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_fees

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_free_trial_conversions

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_new_instances

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_new_product_subscribers

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_canceled_product_subscribers

From 2017-09-15 to present: Available daily by 24:00 UTC.

monthly_revenue_billing_and_revenue_data

From 2017-09-15 to present: Available monthly on the 15th day of the month by 24:00 UTC. Data includes metered transactions (e.g. hourly) from one month prior.

monthly_revenue_annual_subscriptions

From 2017-09-15 to present: Available monthly on the 15th day of the month by 24:00 UTC. Data includes up-front software charges (e.g. annual) from one month prior.

monthly_revenue_field_demonstration_usage

From 2018-03-15 to present: Available monthly on the 15th day of the month by 24:00 UTC.

monthly_revenue_flexible_payment_schedule

From 2018-11-15 to present: Available monthly on the 15th day of the month by 24:00 UTC.

disbursed_amount_by_product

From 2017-09-15 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_instance_hours

From 2017-09-15 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_customer_geo

From 2017-09-15 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_age_of_uncollected_funds

From 2017-09-15 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_age_of_disbursed_funds

From 2017-09-15 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_age_of_past_due_funds

From 2018-04-07 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_uncollected_funds_breakdown

From 2019-10-04 to present: Available every 30 days by 24:00 UTC.

sales_compensation_billed_revenue

From 2017-09-15 to present: Available monthly on the 15th day of the month by 24:00 UTC. Data includes metered transactions (e.g. hourly) from one month prior, and up-front software charges (e.g. annual) from one month prior.

us_sales_and_use_tax_records

From 2017-09-15 to present: Available monthly on the 15th day of the month by 24:00 UTC.

disbursed_amount_by_product_with_uncollected_funds

This data set is deprecated. Download related reports from AMMP instead!

customer_profile_by_industry

This data set is deprecated. Download related reports from AMMP instead!

customer_profile_by_revenue

This data set is deprecated. Download related reports from AMMP instead!

customer_profile_by_geography

This data set is deprecated. Download related reports from AMMP instead!

$sel:dataSetPublicationDate:GenerateDataSet', generateDataSet_dataSetPublicationDate - The date a data set was published. For daily data sets, provide a date with day-level granularity for the desired day. For monthly data sets except those with prefix disbursed_amount, provide a date with month-level granularity for the desired month (the day value will be ignored). For data sets with prefix disbursed_amount, provide a date with day-level granularity for the desired day. For these data sets we will look backwards in time over the range of 31 days until the first data set is found (the latest one).

$sel:roleNameArn:GenerateDataSet', generateDataSet_roleNameArn - The Amazon Resource Name (ARN) of the Role with an attached permissions policy to interact with the provided AWS services.

$sel:destinationS3BucketName:GenerateDataSet', generateDataSet_destinationS3BucketName - The name (friendly name, not ARN) of the destination S3 bucket.

$sel:snsTopicArn:GenerateDataSet', generateDataSet_snsTopicArn - Amazon Resource Name (ARN) for the SNS Topic that will be notified when the data set has been published or if an error has occurred.

Request Lenses

generateDataSet_destinationS3Prefix :: Lens' GenerateDataSet (Maybe Text)

(Optional) The desired S3 prefix for the published data set, similar to a directory path in standard file systems. For example, if given the bucket name "mybucket" and the prefix "myprefix/mydatasets", the output file "outputfile" would be published to "s3://mybucket/myprefix/mydatasets/outputfile". If the prefix directory structure does not exist, it will be created. If no prefix is provided, the data set will be published to the S3 bucket root.

generateDataSet_customerDefinedValues :: Lens' GenerateDataSet (Maybe (HashMap Text Text))

(Optional) Key-value pairs which will be returned, unmodified, in the Amazon SNS notification message and the data set metadata file. These key-value pairs can be used to correlated responses with tracking information from other systems.

generateDataSet_dataSetType :: Lens' GenerateDataSet DataSetType

The desired data set type.

customer_subscriber_hourly_monthly_subscriptions

From 2017-09-15 to present: Available daily by 24:00 UTC.

customer_subscriber_annual_subscriptions

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_usage_by_instance_type

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_fees

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_free_trial_conversions

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_new_instances

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_new_product_subscribers

From 2017-09-15 to present: Available daily by 24:00 UTC.

daily_business_canceled_product_subscribers

From 2017-09-15 to present: Available daily by 24:00 UTC.

monthly_revenue_billing_and_revenue_data

From 2017-09-15 to present: Available monthly on the 15th day of the month by 24:00 UTC. Data includes metered transactions (e.g. hourly) from one month prior.

monthly_revenue_annual_subscriptions

From 2017-09-15 to present: Available monthly on the 15th day of the month by 24:00 UTC. Data includes up-front software charges (e.g. annual) from one month prior.

monthly_revenue_field_demonstration_usage

From 2018-03-15 to present: Available monthly on the 15th day of the month by 24:00 UTC.

monthly_revenue_flexible_payment_schedule

From 2018-11-15 to present: Available monthly on the 15th day of the month by 24:00 UTC.

disbursed_amount_by_product

From 2017-09-15 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_instance_hours

From 2017-09-15 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_customer_geo

From 2017-09-15 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_age_of_uncollected_funds

From 2017-09-15 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_age_of_disbursed_funds

From 2017-09-15 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_age_of_past_due_funds

From 2018-04-07 to present: Available every 30 days by 24:00 UTC.

disbursed_amount_by_uncollected_funds_breakdown

From 2019-10-04 to present: Available every 30 days by 24:00 UTC.

sales_compensation_billed_revenue

From 2017-09-15 to present: Available monthly on the 15th day of the month by 24:00 UTC. Data includes metered transactions (e.g. hourly) from one month prior, and up-front software charges (e.g. annual) from one month prior.

us_sales_and_use_tax_records

From 2017-09-15 to present: Available monthly on the 15th day of the month by 24:00 UTC.

disbursed_amount_by_product_with_uncollected_funds

This data set is deprecated. Download related reports from AMMP instead!

customer_profile_by_industry

This data set is deprecated. Download related reports from AMMP instead!

customer_profile_by_revenue

This data set is deprecated. Download related reports from AMMP instead!

customer_profile_by_geography

This data set is deprecated. Download related reports from AMMP instead!

generateDataSet_dataSetPublicationDate :: Lens' GenerateDataSet UTCTime

The date a data set was published. For daily data sets, provide a date with day-level granularity for the desired day. For monthly data sets except those with prefix disbursed_amount, provide a date with month-level granularity for the desired month (the day value will be ignored). For data sets with prefix disbursed_amount, provide a date with day-level granularity for the desired day. For these data sets we will look backwards in time over the range of 31 days until the first data set is found (the latest one).

generateDataSet_roleNameArn :: Lens' GenerateDataSet Text

The Amazon Resource Name (ARN) of the Role with an attached permissions policy to interact with the provided AWS services.

generateDataSet_destinationS3BucketName :: Lens' GenerateDataSet Text

The name (friendly name, not ARN) of the destination S3 bucket.

generateDataSet_snsTopicArn :: Lens' GenerateDataSet Text

Amazon Resource Name (ARN) for the SNS Topic that will be notified when the data set has been published or if an error has occurred.

Destructuring the Response

data GenerateDataSetResponse

Container for the result of the GenerateDataSet operation.

See: newGenerateDataSetResponse smart constructor.

Constructors

GenerateDataSetResponse'

Fields

dataSetRequestId :: Maybe Text

A unique identifier representing a specific request to the GenerateDataSet operation. This identifier can be used to correlate a request with notifications from the SNS topic.

httpStatus :: Int

The response's http status code.

newGenerateDataSetResponse

Arguments

:: Int
$sel:httpStatus:GenerateDataSetResponse'

-> GenerateDataSetResponse

Create a value of GenerateDataSetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataSetRequestId:GenerateDataSetResponse', generateDataSetResponse_dataSetRequestId - A unique identifier representing a specific request to the GenerateDataSet operation. This identifier can be used to correlate a request with notifications from the SNS topic.

$sel:httpStatus:GenerateDataSetResponse', generateDataSetResponse_httpStatus - The response's http status code.

Response Lenses

generateDataSetResponse_dataSetRequestId :: Lens' GenerateDataSetResponse (Maybe Text)

A unique identifier representing a specific request to the GenerateDataSet operation. This identifier can be used to correlate a request with notifications from the SNS topic.

generateDataSetResponse_httpStatus :: Lens' GenerateDataSetResponse Int

The response's http status code.