Network.AWS.MarketplaceAnalytics.StartSupportDataExport

Description

Given a data set type and a from date, asynchronously publishes the requested customer support data to the specified S3 bucket and notifies the specified SNS topic once the data is available. Returns a unique request identifier that can be used to correlate requests with notifications from the SNS topic. Data sets will be published in comma-separated values (CSV) format with the file name {data_set_type}_YYYY-MM-DD'T'HH-mm-ss'Z'.csv. If a file with the same name already exists (e.g. if the same data set is requested twice), the original file will be overwritten by the new file. Requires a Role with an attached permissions policy providing Allow permissions for the following actions: s3:PutObject, s3:GetBucketLocation, sns:GetTopicAttributes, sns:Publish, iam:GetRolePolicy.

Creating a Request

data StartSupportDataExport

Container for the parameters to the StartSupportDataExport operation.

See: newStartSupportDataExport smart constructor.

Constructors

StartSupportDataExport'

Fields

destinationS3Prefix :: Maybe Text

(Optional) The desired S3 prefix for the published data set, similar to a directory path in standard file systems. For example, if given the bucket name "mybucket" and the prefix "myprefix/mydatasets", the output file "outputfile" would be published to "s3://mybucket/myprefix/mydatasets/outputfile". If the prefix directory structure does not exist, it will be created. If no prefix is provided, the data set will be published to the S3 bucket root.

customerDefinedValues :: Maybe (HashMap Text Text)

(Optional) Key-value pairs which will be returned, unmodified, in the Amazon SNS notification message and the data set metadata file.

dataSetType :: SupportDataSetType

Specifies the data set type to be written to the output csv file. The data set types customer_support_contacts_data and test_customer_support_contacts_data both result in a csv file containing the following fields: Product Id, Product Code, Customer Guid, Subscription Guid, Subscription Start Date, Organization, AWS Account Id, Given Name, Surname, Telephone Number, Email, Title, Country Code, ZIP Code, Operation Type, and Operation Time.

customer_support_contacts_data Customer support contact data. The data set will contain all changes (Creates, Updates, and Deletes) to customer support contact data from the date specified in the from_date parameter.

test_customer_support_contacts_data An example data set containing static test data in the same format as customer_support_contacts_data

fromDate :: POSIX

The start date from which to retrieve the data set in UTC. This parameter only affects the customer_support_contacts_data data set type.

roleNameArn :: Text

The Amazon Resource Name (ARN) of the Role with an attached permissions policy to interact with the provided AWS services.

destinationS3BucketName :: Text

The name (friendly name, not ARN) of the destination S3 bucket.

snsTopicArn :: Text

Amazon Resource Name (ARN) for the SNS Topic that will be notified when the data set has been published or if an error has occurred.

newStartSupportDataExport

Arguments

:: SupportDataSetType
$sel:dataSetType:StartSupportDataExport'

-> UTCTime
$sel:fromDate:StartSupportDataExport'

-> Text
$sel:roleNameArn:StartSupportDataExport'

-> Text
$sel:destinationS3BucketName:StartSupportDataExport'

-> Text
$sel:snsTopicArn:StartSupportDataExport'

-> StartSupportDataExport

Create a value of StartSupportDataExport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationS3Prefix:StartSupportDataExport', startSupportDataExport_destinationS3Prefix - (Optional) The desired S3 prefix for the published data set, similar to a directory path in standard file systems. For example, if given the bucket name "mybucket" and the prefix "myprefix/mydatasets", the output file "outputfile" would be published to "s3://mybucket/myprefix/mydatasets/outputfile". If the prefix directory structure does not exist, it will be created. If no prefix is provided, the data set will be published to the S3 bucket root.

$sel:customerDefinedValues:StartSupportDataExport', startSupportDataExport_customerDefinedValues - (Optional) Key-value pairs which will be returned, unmodified, in the Amazon SNS notification message and the data set metadata file.

$sel:dataSetType:StartSupportDataExport', startSupportDataExport_dataSetType - Specifies the data set type to be written to the output csv file. The data set types customer_support_contacts_data and test_customer_support_contacts_data both result in a csv file containing the following fields: Product Id, Product Code, Customer Guid, Subscription Guid, Subscription Start Date, Organization, AWS Account Id, Given Name, Surname, Telephone Number, Email, Title, Country Code, ZIP Code, Operation Type, and Operation Time.

customer_support_contacts_data Customer support contact data. The data set will contain all changes (Creates, Updates, and Deletes) to customer support contact data from the date specified in the from_date parameter.

test_customer_support_contacts_data An example data set containing static test data in the same format as customer_support_contacts_data

$sel:fromDate:StartSupportDataExport', startSupportDataExport_fromDate - The start date from which to retrieve the data set in UTC. This parameter only affects the customer_support_contacts_data data set type.

$sel:roleNameArn:StartSupportDataExport', startSupportDataExport_roleNameArn - The Amazon Resource Name (ARN) of the Role with an attached permissions policy to interact with the provided AWS services.

$sel:destinationS3BucketName:StartSupportDataExport', startSupportDataExport_destinationS3BucketName - The name (friendly name, not ARN) of the destination S3 bucket.

$sel:snsTopicArn:StartSupportDataExport', startSupportDataExport_snsTopicArn - Amazon Resource Name (ARN) for the SNS Topic that will be notified when the data set has been published or if an error has occurred.

Request Lenses

startSupportDataExport_destinationS3Prefix :: Lens' StartSupportDataExport (Maybe Text)

(Optional) The desired S3 prefix for the published data set, similar to a directory path in standard file systems. For example, if given the bucket name "mybucket" and the prefix "myprefix/mydatasets", the output file "outputfile" would be published to "s3://mybucket/myprefix/mydatasets/outputfile". If the prefix directory structure does not exist, it will be created. If no prefix is provided, the data set will be published to the S3 bucket root.

startSupportDataExport_customerDefinedValues :: Lens' StartSupportDataExport (Maybe (HashMap Text Text))

(Optional) Key-value pairs which will be returned, unmodified, in the Amazon SNS notification message and the data set metadata file.

startSupportDataExport_dataSetType :: Lens' StartSupportDataExport SupportDataSetType

Specifies the data set type to be written to the output csv file. The data set types customer_support_contacts_data and test_customer_support_contacts_data both result in a csv file containing the following fields: Product Id, Product Code, Customer Guid, Subscription Guid, Subscription Start Date, Organization, AWS Account Id, Given Name, Surname, Telephone Number, Email, Title, Country Code, ZIP Code, Operation Type, and Operation Time.

customer_support_contacts_data Customer support contact data. The data set will contain all changes (Creates, Updates, and Deletes) to customer support contact data from the date specified in the from_date parameter.

test_customer_support_contacts_data An example data set containing static test data in the same format as customer_support_contacts_data

startSupportDataExport_fromDate :: Lens' StartSupportDataExport UTCTime

The start date from which to retrieve the data set in UTC. This parameter only affects the customer_support_contacts_data data set type.

startSupportDataExport_roleNameArn :: Lens' StartSupportDataExport Text

The Amazon Resource Name (ARN) of the Role with an attached permissions policy to interact with the provided AWS services.

startSupportDataExport_destinationS3BucketName :: Lens' StartSupportDataExport Text

The name (friendly name, not ARN) of the destination S3 bucket.

startSupportDataExport_snsTopicArn :: Lens' StartSupportDataExport Text

Amazon Resource Name (ARN) for the SNS Topic that will be notified when the data set has been published or if an error has occurred.

Destructuring the Response

data StartSupportDataExportResponse

Container for the result of the StartSupportDataExport operation.

See: newStartSupportDataExportResponse smart constructor.

Constructors

StartSupportDataExportResponse'

Fields

dataSetRequestId :: Maybe Text

A unique identifier representing a specific request to the StartSupportDataExport operation. This identifier can be used to correlate a request with notifications from the SNS topic.

httpStatus :: Int

The response's http status code.

newStartSupportDataExportResponse

Arguments

:: Int
$sel:httpStatus:StartSupportDataExportResponse'

-> StartSupportDataExportResponse

Create a value of StartSupportDataExportResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataSetRequestId:StartSupportDataExportResponse', startSupportDataExportResponse_dataSetRequestId - A unique identifier representing a specific request to the StartSupportDataExport operation. This identifier can be used to correlate a request with notifications from the SNS topic.

$sel:httpStatus:StartSupportDataExportResponse', startSupportDataExportResponse_httpStatus - The response's http status code.

Response Lenses

startSupportDataExportResponse_dataSetRequestId :: Lens' StartSupportDataExportResponse (Maybe Text)

A unique identifier representing a specific request to the StartSupportDataExport operation. This identifier can be used to correlate a request with notifications from the SNS topic.

startSupportDataExportResponse_httpStatus :: Lens' StartSupportDataExportResponse Int

The response's http status code.