Network.AWS.MarketplaceCatalog.Types.EntitySummary

Description

Documentation

data EntitySummary

This object is a container for common summary information about the entity. The summary doesn't contain the whole entity structure, but it does contain information common across all entities.

See: newEntitySummary smart constructor.

Constructors

EntitySummary'

Fields

lastModifiedDate :: Maybe Text

The last time the entity was published, using ISO 8601 format (2018-02-27T13:45:22Z).

entityType :: Maybe Text

The type of the entity.

visibility :: Maybe Text

The visibility status of the entity to buyers. This value can be Public (everyone can view the entity), Limited (the entity is visible to limited accounts only), or Restricted (the entity was published and then unpublished and only existing buyers can view it).

name :: Maybe Text

The name for the entity. This value is not unique. It is defined by the seller.

entityId :: Maybe Text

The unique identifier for the entity.

entityArn :: Maybe Text

The ARN associated with the unique identifier for the entity.

newEntitySummary :: EntitySummary

Create a value of EntitySummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:EntitySummary', entitySummary_lastModifiedDate - The last time the entity was published, using ISO 8601 format (2018-02-27T13:45:22Z).

$sel:entityType:EntitySummary', entitySummary_entityType - The type of the entity.

$sel:visibility:EntitySummary', entitySummary_visibility - The visibility status of the entity to buyers. This value can be Public (everyone can view the entity), Limited (the entity is visible to limited accounts only), or Restricted (the entity was published and then unpublished and only existing buyers can view it).

$sel:name:EntitySummary', entitySummary_name - The name for the entity. This value is not unique. It is defined by the seller.

$sel:entityId:EntitySummary', entitySummary_entityId - The unique identifier for the entity.

$sel:entityArn:EntitySummary', entitySummary_entityArn - The ARN associated with the unique identifier for the entity.

entitySummary_lastModifiedDate :: Lens' EntitySummary (Maybe Text)

The last time the entity was published, using ISO 8601 format (2018-02-27T13:45:22Z).

entitySummary_entityType :: Lens' EntitySummary (Maybe Text)

The type of the entity.

entitySummary_visibility :: Lens' EntitySummary (Maybe Text)

The visibility status of the entity to buyers. This value can be Public (everyone can view the entity), Limited (the entity is visible to limited accounts only), or Restricted (the entity was published and then unpublished and only existing buyers can view it).

entitySummary_name :: Lens' EntitySummary (Maybe Text)

The name for the entity. This value is not unique. It is defined by the seller.

entitySummary_entityId :: Lens' EntitySummary (Maybe Text)

The unique identifier for the entity.

entitySummary_entityArn :: Lens' EntitySummary (Maybe Text)

The ARN associated with the unique identifier for the entity.