Network.AWS.MarketplaceEntitlement.Lens

Description

Operations

GetEntitlements

getEntitlements_nextToken :: Lens' GetEntitlements (Maybe Text)

For paginated calls to GetEntitlements, pass the NextToken from the previous GetEntitlementsResult.

getEntitlements_maxResults :: Lens' GetEntitlements (Maybe Int)

The maximum number of items to retrieve from the GetEntitlements operation. For pagination, use the NextToken field in subsequent calls to GetEntitlements.

getEntitlements_filter :: Lens' GetEntitlements (Maybe (HashMap GetEntitlementFilterName (NonEmpty Text)))

Filter is used to return entitlements for a specific customer or for a specific dimension. Filters are described as keys mapped to a lists of values. Filtered requests are unioned for each value in the value list, and then intersected for each filter key.

getEntitlements_productCode :: Lens' GetEntitlements Text

Product code is used to uniquely identify a product in AWS Marketplace. The product code will be provided by AWS Marketplace when the product listing is created.

getEntitlementsResponse_nextToken :: Lens' GetEntitlementsResponse (Maybe Text)

For paginated results, use NextToken in subsequent calls to GetEntitlements. If the result contains an empty set of entitlements, NextToken might still be present and should be used.

getEntitlementsResponse_entitlements :: Lens' GetEntitlementsResponse (Maybe [Entitlement])

The set of entitlements found through the GetEntitlements operation. If the result contains an empty set of entitlements, NextToken might still be present and should be used.

getEntitlementsResponse_httpStatus :: Lens' GetEntitlementsResponse Int

The response's http status code.

Types

Entitlement

entitlement_expirationDate :: Lens' Entitlement (Maybe UTCTime)

The expiration date represents the minimum date through which this entitlement is expected to remain valid. For contractual products listed on AWS Marketplace, the expiration date is the date at which the customer will renew or cancel their contract. Customers who are opting to renew their contract will still have entitlements with an expiration date.

entitlement_customerIdentifier :: Lens' Entitlement (Maybe Text)

The customer identifier is a handle to each unique customer in an application. Customer identifiers are obtained through the ResolveCustomer operation in AWS Marketplace Metering Service.

entitlement_productCode :: Lens' Entitlement (Maybe Text)

The product code for which the given entitlement applies. Product codes are provided by AWS Marketplace when the product listing is created.

entitlement_value :: Lens' Entitlement (Maybe EntitlementValue)

The EntitlementValue represents the amount of capacity that the customer is entitled to for the product.

entitlement_dimension :: Lens' Entitlement (Maybe Text)

The dimension for which the given entitlement applies. Dimensions represent categories of capacity in a product and are specified when the product is listed in AWS Marketplace.

EntitlementValue

entitlementValue_doubleValue :: Lens' EntitlementValue (Maybe Double)

The DoubleValue field will be populated with a double value when the entitlement is a double type. Otherwise, the field will not be set.

entitlementValue_stringValue :: Lens' EntitlementValue (Maybe Text)

The StringValue field will be populated with a string value when the entitlement is a string type. Otherwise, the field will not be set.

entitlementValue_booleanValue :: Lens' EntitlementValue (Maybe Bool)

The BooleanValue field will be populated with a boolean value when the entitlement is a boolean type. Otherwise, the field will not be set.

entitlementValue_integerValue :: Lens' EntitlementValue (Maybe Int)

The IntegerValue field will be populated with an integer value when the entitlement is an integer type. Otherwise, the field will not be set.