Network.AWS.MarketplaceMetering.MeterUsage

Description

API to emit metering records. For identical requests, the API is idempotent. It simply returns the metering record ID.

MeterUsage is authenticated on the buyer's AWS account using credentials from the EC2 instance, ECS task, or EKS pod.

MeterUsage can optionally include multiple usage allocations, to provide customers with usage data split into buckets by tags that you define (or allow the customer to define).

Creating a Request

data MeterUsage

See: newMeterUsage smart constructor.

Constructors

MeterUsage'

Fields

usageAllocations :: Maybe (NonEmpty UsageAllocation)

The set of UsageAllocations to submit.

The sum of all UsageAllocation quantities must equal the UsageQuantity of the MeterUsage request, and each UsageAllocation must have a unique set of tags (include no tags).

usageQuantity :: Maybe Natural

Consumption value for the hour. Defaults to 0 if not specified.

dryRun :: Maybe Bool

Checks whether you have the permissions required for the action, but does not make the request. If you have the permissions, the request returns DryRunOperation; otherwise, it returns UnauthorizedException. Defaults to false if not specified.

productCode :: Text

Product code is used to uniquely identify a product in AWS Marketplace. The product code should be the same as the one used during the publishing of a new product.

timestamp :: POSIX

Timestamp, in UTC, for which the usage is being reported. Your application can meter usage for up to one hour in the past. Make sure the timestamp value is not before the start of the software usage.

usageDimension :: Text

It will be one of the fcp dimension name provided during the publishing of the product.

newMeterUsage

Arguments

:: Text
$sel:productCode:MeterUsage'

-> UTCTime
$sel:timestamp:MeterUsage'

-> Text
$sel:usageDimension:MeterUsage'

-> MeterUsage

Create a value of MeterUsage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:usageAllocations:MeterUsage', meterUsage_usageAllocations - The set of UsageAllocations to submit.

The sum of all UsageAllocation quantities must equal the UsageQuantity of the MeterUsage request, and each UsageAllocation must have a unique set of tags (include no tags).

$sel:usageQuantity:MeterUsage', meterUsage_usageQuantity - Consumption value for the hour. Defaults to 0 if not specified.

$sel:dryRun:MeterUsage', meterUsage_dryRun - Checks whether you have the permissions required for the action, but does not make the request. If you have the permissions, the request returns DryRunOperation; otherwise, it returns UnauthorizedException. Defaults to false if not specified.

$sel:productCode:MeterUsage', meterUsage_productCode - Product code is used to uniquely identify a product in AWS Marketplace. The product code should be the same as the one used during the publishing of a new product.

$sel:timestamp:MeterUsage', meterUsage_timestamp - Timestamp, in UTC, for which the usage is being reported. Your application can meter usage for up to one hour in the past. Make sure the timestamp value is not before the start of the software usage.

$sel:usageDimension:MeterUsage', meterUsage_usageDimension - It will be one of the fcp dimension name provided during the publishing of the product.

Request Lenses

meterUsage_usageAllocations :: Lens' MeterUsage (Maybe (NonEmpty UsageAllocation))

The set of UsageAllocations to submit.

The sum of all UsageAllocation quantities must equal the UsageQuantity of the MeterUsage request, and each UsageAllocation must have a unique set of tags (include no tags).

meterUsage_usageQuantity :: Lens' MeterUsage (Maybe Natural)

Consumption value for the hour. Defaults to 0 if not specified.

meterUsage_dryRun :: Lens' MeterUsage (Maybe Bool)

Checks whether you have the permissions required for the action, but does not make the request. If you have the permissions, the request returns DryRunOperation; otherwise, it returns UnauthorizedException. Defaults to false if not specified.

meterUsage_productCode :: Lens' MeterUsage Text

Product code is used to uniquely identify a product in AWS Marketplace. The product code should be the same as the one used during the publishing of a new product.

meterUsage_timestamp :: Lens' MeterUsage UTCTime

Timestamp, in UTC, for which the usage is being reported. Your application can meter usage for up to one hour in the past. Make sure the timestamp value is not before the start of the software usage.

meterUsage_usageDimension :: Lens' MeterUsage Text

It will be one of the fcp dimension name provided during the publishing of the product.

Destructuring the Response

data MeterUsageResponse

See: newMeterUsageResponse smart constructor.

Constructors

MeterUsageResponse'

Fields

meteringRecordId :: Maybe Text

Metering record id.

httpStatus :: Int

The response's http status code.

newMeterUsageResponse

Arguments

:: Int
$sel:httpStatus:MeterUsageResponse'

-> MeterUsageResponse

Create a value of MeterUsageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:meteringRecordId:MeterUsageResponse', meterUsageResponse_meteringRecordId - Metering record id.

$sel:httpStatus:MeterUsageResponse', meterUsageResponse_httpStatus - The response's http status code.

Response Lenses

meterUsageResponse_meteringRecordId :: Lens' MeterUsageResponse (Maybe Text)

Metering record id.

meterUsageResponse_httpStatus :: Lens' MeterUsageResponse Int

The response's http status code.