Network.AWS.MarketplaceMetering.ResolveCustomer

Description

ResolveCustomer is called by a SaaS application during the registration process. When a buyer visits your website during the registration process, the buyer submits a registration token through their browser. The registration token is resolved through this API to obtain a CustomerIdentifier and product code.

Creating a Request

data ResolveCustomer

Contains input to the ResolveCustomer operation.

See: newResolveCustomer smart constructor.

Constructors

ResolveCustomer'

Fields

registrationToken :: Text

When a buyer visits your website during the registration process, the buyer submits a registration token through the browser. The registration token is resolved to obtain a CustomerIdentifier and product code.

newResolveCustomer

Arguments

:: Text
$sel:registrationToken:ResolveCustomer'

-> ResolveCustomer

Create a value of ResolveCustomer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:registrationToken:ResolveCustomer', resolveCustomer_registrationToken - When a buyer visits your website during the registration process, the buyer submits a registration token through the browser. The registration token is resolved to obtain a CustomerIdentifier and product code.

Request Lenses

resolveCustomer_registrationToken :: Lens' ResolveCustomer Text

When a buyer visits your website during the registration process, the buyer submits a registration token through the browser. The registration token is resolved to obtain a CustomerIdentifier and product code.

Destructuring the Response

data ResolveCustomerResponse

The result of the ResolveCustomer operation. Contains the CustomerIdentifier and product code.

See: newResolveCustomerResponse smart constructor.

Constructors

ResolveCustomerResponse'

Fields

customerIdentifier :: Maybe Text

The CustomerIdentifier is used to identify an individual customer in your application. Calls to BatchMeterUsage require CustomerIdentifiers for each UsageRecord.

productCode :: Maybe Text

The product code is returned to confirm that the buyer is registering for your product. Subsequent BatchMeterUsage calls should be made using this product code.

httpStatus :: Int

The response's http status code.

newResolveCustomerResponse

Arguments

:: Int
$sel:httpStatus:ResolveCustomerResponse'

-> ResolveCustomerResponse

Create a value of ResolveCustomerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customerIdentifier:ResolveCustomerResponse', resolveCustomerResponse_customerIdentifier - The CustomerIdentifier is used to identify an individual customer in your application. Calls to BatchMeterUsage require CustomerIdentifiers for each UsageRecord.

$sel:productCode:ResolveCustomerResponse', resolveCustomerResponse_productCode - The product code is returned to confirm that the buyer is registering for your product. Subsequent BatchMeterUsage calls should be made using this product code.

$sel:httpStatus:ResolveCustomerResponse', resolveCustomerResponse_httpStatus - The response's http status code.

Response Lenses

resolveCustomerResponse_customerIdentifier :: Lens' ResolveCustomerResponse (Maybe Text)

The CustomerIdentifier is used to identify an individual customer in your application. Calls to BatchMeterUsage require CustomerIdentifiers for each UsageRecord.

resolveCustomerResponse_productCode :: Lens' ResolveCustomerResponse (Maybe Text)

The product code is returned to confirm that the buyer is registering for your product. Subsequent BatchMeterUsage calls should be made using this product code.

resolveCustomerResponse_httpStatus :: Lens' ResolveCustomerResponse Int

The response's http status code.