Network.AWS.MarketplaceMetering.Types

Description

Service Configuration

defaultService :: Service

API version 2016-01-14 of the Amazon Marketplace Metering SDK configuration.

Errors

_TimestampOutOfBoundsException :: AsError a => Getting (First ServiceError) a ServiceError

The timestamp value passed in the meterUsage() is out of allowed range.

_CustomerNotEntitledException :: AsError a => Getting (First ServiceError) a ServiceError

Exception thrown when the customer does not have a valid subscription for the product.

_InvalidRegionException :: AsError a => Getting (First ServiceError) a ServiceError

RegisterUsage must be called in the same AWS Region the ECS task was launched in. This prevents a container from hardcoding a Region (e.g. withRegion(“us-east-1”) when calling RegisterUsage.

_ExpiredTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The submitted registration token has expired. This can happen if the buyer's browser takes too long to redirect to your page, the buyer has resubmitted the registration token, or your application has held on to the registration token for too long. Your SaaS registration website should redeem this token as soon as it is submitted by the buyer's browser.

_InvalidUsageDimensionException :: AsError a => Getting (First ServiceError) a ServiceError

The usage dimension does not match one of the UsageDimensions associated with products.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The calls to the API are throttled.

_DisabledApiException :: AsError a => Getting (First ServiceError) a ServiceError

The API is disabled in the Region.

_InvalidTagException :: AsError a => Getting (First ServiceError) a ServiceError

The tag is invalid, or the number of tags is greater than 5.

_InvalidCustomerIdentifierException :: AsError a => Getting (First ServiceError) a ServiceError

You have metered usage for a CustomerIdentifier that does not exist.

_DuplicateRequestException :: AsError a => Getting (First ServiceError) a ServiceError

A metering record has already been emitted by the same EC2 instance, ECS task, or EKS pod for the given {usageDimension, timestamp} with a different usageQuantity.

_PlatformNotSupportedException :: AsError a => Getting (First ServiceError) a ServiceError

AWS Marketplace does not support metering usage from the underlying platform. Currently, Amazon ECS, Amazon EKS, and AWS Fargate are supported.

_InvalidProductCodeException :: AsError a => Getting (First ServiceError) a ServiceError

The product code passed does not match the product code used for publishing the product.

_InvalidUsageAllocationsException :: AsError a => Getting (First ServiceError) a ServiceError

The usage allocation objects are invalid, or the number of allocations is greater than 500 for a single usage record.

_InternalServiceErrorException :: AsError a => Getting (First ServiceError) a ServiceError

An internal error has occurred. Retry your request. If the problem persists, post a message with details on the AWS forums.

_InvalidTokenException :: AsError a => Getting (First ServiceError) a ServiceError

Registration token is invalid.

_InvalidEndpointRegionException :: AsError a => Getting (First ServiceError) a ServiceError

The endpoint being called is in a AWS Region different from your EC2 instance, ECS task, or EKS pod. The Region of the Metering Service endpoint and the AWS Region of the resource must match.

_InvalidPublicKeyVersionException :: AsError a => Getting (First ServiceError) a ServiceError

Public Key version is invalid.

UsageRecordResultStatus

newtype UsageRecordResultStatus

Constructors

UsageRecordResultStatus'

Fields

fromUsageRecordResultStatus :: Text

Bundled Patterns

pattern UsageRecordResultStatus_CustomerNotSubscribed :: UsageRecordResultStatus

pattern UsageRecordResultStatus_DuplicateRecord :: UsageRecordResultStatus

pattern UsageRecordResultStatus_Success :: UsageRecordResultStatus

Tag

data Tag

Metadata assigned to an allocation. Each tag is made up of a key and a value.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Text

One part of a key-value pair that makes up a tag. A key is a label that acts like a category for the specific tag values.

value :: Text

One part of a key-value pair that makes up a tag. A value acts as a descriptor within a tag category (key). The value can be empty or null.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - One part of a key-value pair that makes up a tag. A key is a label that acts like a category for the specific tag values.

$sel:value:Tag', tag_value - One part of a key-value pair that makes up a tag. A value acts as a descriptor within a tag category (key). The value can be empty or null.

tag_key :: Lens' Tag Text

One part of a key-value pair that makes up a tag. A key is a label that acts like a category for the specific tag values.

tag_value :: Lens' Tag Text

One part of a key-value pair that makes up a tag. A value acts as a descriptor within a tag category (key). The value can be empty or null.

UsageAllocation

data UsageAllocation

Usage allocations allow you to split usage into buckets by tags.

Each UsageAllocation indicates the usage quantity for a specific set of tags.

See: newUsageAllocation smart constructor.

Constructors

UsageAllocation'

Fields

tags :: Maybe (NonEmpty Tag)

The set of tags that define the bucket of usage. For the bucket of items with no tags, this parameter can be left out.

allocatedUsageQuantity :: Natural

The total quantity allocated to this bucket of usage.

newUsageAllocation

Arguments

:: Natural
$sel:allocatedUsageQuantity:UsageAllocation'

-> UsageAllocation

Create a value of UsageAllocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:UsageAllocation', usageAllocation_tags - The set of tags that define the bucket of usage. For the bucket of items with no tags, this parameter can be left out.

$sel:allocatedUsageQuantity:UsageAllocation', usageAllocation_allocatedUsageQuantity - The total quantity allocated to this bucket of usage.

usageAllocation_tags :: Lens' UsageAllocation (Maybe (NonEmpty Tag))

The set of tags that define the bucket of usage. For the bucket of items with no tags, this parameter can be left out.

usageAllocation_allocatedUsageQuantity :: Lens' UsageAllocation Natural

The total quantity allocated to this bucket of usage.

UsageRecord

data UsageRecord

A UsageRecord indicates a quantity of usage for a given product, customer, dimension and time.

Multiple requests with the same UsageRecords as input will be deduplicated to prevent double charges.

See: newUsageRecord smart constructor.

Constructors

UsageRecord'

Fields

usageAllocations :: Maybe (NonEmpty UsageAllocation)

The set of UsageAllocations to submit. The sum of all UsageAllocation quantities must equal the Quantity of the UsageRecord.

quantity :: Maybe Natural

The quantity of usage consumed by the customer for the given dimension and time. Defaults to 0 if not specified.

timestamp :: POSIX

Timestamp, in UTC, for which the usage is being reported.

Your application can meter usage for up to one hour in the past. Make sure the timestamp value is not before the start of the software usage.

customerIdentifier :: Text

The CustomerIdentifier is obtained through the ResolveCustomer operation and represents an individual buyer in your application.

dimension :: Text

During the process of registering a product on AWS Marketplace, up to eight dimensions are specified. These represent different units of value in your application.

newUsageRecord

Arguments

:: UTCTime
$sel:timestamp:UsageRecord'

-> Text
$sel:customerIdentifier:UsageRecord'

-> Text
$sel:dimension:UsageRecord'

-> UsageRecord

Create a value of UsageRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:usageAllocations:UsageRecord', usageRecord_usageAllocations - The set of UsageAllocations to submit. The sum of all UsageAllocation quantities must equal the Quantity of the UsageRecord.

$sel:quantity:UsageRecord', usageRecord_quantity - The quantity of usage consumed by the customer for the given dimension and time. Defaults to 0 if not specified.

$sel:timestamp:UsageRecord', usageRecord_timestamp - Timestamp, in UTC, for which the usage is being reported.

Your application can meter usage for up to one hour in the past. Make sure the timestamp value is not before the start of the software usage.

$sel:customerIdentifier:UsageRecord', usageRecord_customerIdentifier - The CustomerIdentifier is obtained through the ResolveCustomer operation and represents an individual buyer in your application.

$sel:dimension:UsageRecord', usageRecord_dimension - During the process of registering a product on AWS Marketplace, up to eight dimensions are specified. These represent different units of value in your application.

usageRecord_usageAllocations :: Lens' UsageRecord (Maybe (NonEmpty UsageAllocation))

The set of UsageAllocations to submit. The sum of all UsageAllocation quantities must equal the Quantity of the UsageRecord.

usageRecord_quantity :: Lens' UsageRecord (Maybe Natural)

The quantity of usage consumed by the customer for the given dimension and time. Defaults to 0 if not specified.

usageRecord_timestamp :: Lens' UsageRecord UTCTime

Timestamp, in UTC, for which the usage is being reported.

Your application can meter usage for up to one hour in the past. Make sure the timestamp value is not before the start of the software usage.

usageRecord_customerIdentifier :: Lens' UsageRecord Text

The CustomerIdentifier is obtained through the ResolveCustomer operation and represents an individual buyer in your application.

usageRecord_dimension :: Lens' UsageRecord Text

During the process of registering a product on AWS Marketplace, up to eight dimensions are specified. These represent different units of value in your application.

UsageRecordResult

data UsageRecordResult

A UsageRecordResult indicates the status of a given UsageRecord processed by BatchMeterUsage.

See: newUsageRecordResult smart constructor.

Constructors

UsageRecordResult'

Fields

status :: Maybe UsageRecordResultStatus

The UsageRecordResult Status indicates the status of an individual UsageRecord processed by BatchMeterUsage.

Success- The UsageRecord was accepted and honored by BatchMeterUsage.

CustomerNotSubscribed- The CustomerIdentifier specified is not subscribed to your product. The UsageRecord was not honored. Future UsageRecords for this customer will fail until the customer subscribes to your product.

DuplicateRecord- Indicates that the UsageRecord was invalid and not honored. A previously metered UsageRecord had the same customer, dimension, and time, but a different quantity.

meteringRecordId :: Maybe Text

The MeteringRecordId is a unique identifier for this metering event.

usageRecord :: Maybe UsageRecord

The UsageRecord that was part of the BatchMeterUsage request.

newUsageRecordResult :: UsageRecordResult

Create a value of UsageRecordResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UsageRecordResult', usageRecordResult_status - The UsageRecordResult Status indicates the status of an individual UsageRecord processed by BatchMeterUsage.

Success- The UsageRecord was accepted and honored by BatchMeterUsage.

CustomerNotSubscribed- The CustomerIdentifier specified is not subscribed to your product. The UsageRecord was not honored. Future UsageRecords for this customer will fail until the customer subscribes to your product.

DuplicateRecord- Indicates that the UsageRecord was invalid and not honored. A previously metered UsageRecord had the same customer, dimension, and time, but a different quantity.

$sel:meteringRecordId:UsageRecordResult', usageRecordResult_meteringRecordId - The MeteringRecordId is a unique identifier for this metering event.

$sel:usageRecord:UsageRecordResult', usageRecordResult_usageRecord - The UsageRecord that was part of the BatchMeterUsage request.

usageRecordResult_status :: Lens' UsageRecordResult (Maybe UsageRecordResultStatus)

The UsageRecordResult Status indicates the status of an individual UsageRecord processed by BatchMeterUsage.

Success- The UsageRecord was accepted and honored by BatchMeterUsage.

CustomerNotSubscribed- The CustomerIdentifier specified is not subscribed to your product. The UsageRecord was not honored. Future UsageRecords for this customer will fail until the customer subscribes to your product.

DuplicateRecord- Indicates that the UsageRecord was invalid and not honored. A previously metered UsageRecord had the same customer, dimension, and time, but a different quantity.

usageRecordResult_meteringRecordId :: Lens' UsageRecordResult (Maybe Text)

The MeteringRecordId is a unique identifier for this metering event.

usageRecordResult_usageRecord :: Lens' UsageRecordResult (Maybe UsageRecord)

The UsageRecord that was part of the BatchMeterUsage request.