Network.AWS.MarketplaceMetering.Types.UsageRecord

Description

Documentation

data UsageRecord

A UsageRecord indicates a quantity of usage for a given product, customer, dimension and time.

Multiple requests with the same UsageRecords as input will be deduplicated to prevent double charges.

See: newUsageRecord smart constructor.

Constructors

UsageRecord'

Fields

quantity :: Maybe Natural

The quantity of usage consumed by the customer for the given dimension and time. Defaults to 0 if not specified.

usageAllocations :: Maybe (NonEmpty UsageAllocation)

The set of UsageAllocations to submit. The sum of all UsageAllocation quantities must equal the Quantity of the UsageRecord.

timestamp :: POSIX

Timestamp, in UTC, for which the usage is being reported.

Your application can meter usage for up to one hour in the past. Make sure the timestamp value is not before the start of the software usage.

customerIdentifier :: Text

The CustomerIdentifier is obtained through the ResolveCustomer operation and represents an individual buyer in your application.

dimension :: Text

During the process of registering a product on AWS Marketplace, up to eight dimensions are specified. These represent different units of value in your application.

newUsageRecord

Arguments

:: UTCTime
$sel:timestamp:UsageRecord'

-> Text
$sel:customerIdentifier:UsageRecord'

-> Text
$sel:dimension:UsageRecord'

-> UsageRecord

Create a value of UsageRecord with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:quantity:UsageRecord', usageRecord_quantity - The quantity of usage consumed by the customer for the given dimension and time. Defaults to 0 if not specified.

$sel:usageAllocations:UsageRecord', usageRecord_usageAllocations - The set of UsageAllocations to submit. The sum of all UsageAllocation quantities must equal the Quantity of the UsageRecord.

$sel:timestamp:UsageRecord', usageRecord_timestamp - Timestamp, in UTC, for which the usage is being reported.

Your application can meter usage for up to one hour in the past. Make sure the timestamp value is not before the start of the software usage.

$sel:customerIdentifier:UsageRecord', usageRecord_customerIdentifier - The CustomerIdentifier is obtained through the ResolveCustomer operation and represents an individual buyer in your application.

$sel:dimension:UsageRecord', usageRecord_dimension - During the process of registering a product on AWS Marketplace, up to eight dimensions are specified. These represent different units of value in your application.

usageRecord_quantity :: Lens' UsageRecord (Maybe Natural)

The quantity of usage consumed by the customer for the given dimension and time. Defaults to 0 if not specified.

usageRecord_usageAllocations :: Lens' UsageRecord (Maybe (NonEmpty UsageAllocation))

The set of UsageAllocations to submit. The sum of all UsageAllocation quantities must equal the Quantity of the UsageRecord.

usageRecord_timestamp :: Lens' UsageRecord UTCTime

Timestamp, in UTC, for which the usage is being reported.

Your application can meter usage for up to one hour in the past. Make sure the timestamp value is not before the start of the software usage.

usageRecord_customerIdentifier :: Lens' UsageRecord Text

The CustomerIdentifier is obtained through the ResolveCustomer operation and represents an individual buyer in your application.

usageRecord_dimension :: Lens' UsageRecord Text

During the process of registering a product on AWS Marketplace, up to eight dimensions are specified. These represent different units of value in your application.