Network.AWS.MechanicalTurk.AcceptQualificationRequest

Description

The AcceptQualificationRequest operation approves a Worker's request for a Qualification.

Only the owner of the Qualification type can grant a Qualification request for that type.

A successful request for the AcceptQualificationRequest operation returns with no errors and an empty body.

Creating a Request

data AcceptQualificationRequest

See: newAcceptQualificationRequest smart constructor.

Constructors

AcceptQualificationRequest'

Fields

integerValue :: Maybe Int

The value of the Qualification. You can omit this value if you are using the presence or absence of the Qualification as the basis for a HIT requirement.

qualificationRequestId :: Text

The ID of the Qualification request, as returned by the GetQualificationRequests operation.

newAcceptQualificationRequest

Arguments

:: Text
$sel:qualificationRequestId:AcceptQualificationRequest'

-> AcceptQualificationRequest

Create a value of AcceptQualificationRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:integerValue:AcceptQualificationRequest', acceptQualificationRequest_integerValue - The value of the Qualification. You can omit this value if you are using the presence or absence of the Qualification as the basis for a HIT requirement.

$sel:qualificationRequestId:AcceptQualificationRequest', acceptQualificationRequest_qualificationRequestId - The ID of the Qualification request, as returned by the GetQualificationRequests operation.

Request Lenses

acceptQualificationRequest_integerValue :: Lens' AcceptQualificationRequest (Maybe Int)

The value of the Qualification. You can omit this value if you are using the presence or absence of the Qualification as the basis for a HIT requirement.

acceptQualificationRequest_qualificationRequestId :: Lens' AcceptQualificationRequest Text

The ID of the Qualification request, as returned by the GetQualificationRequests operation.

Destructuring the Response

data AcceptQualificationRequestResponse

See: newAcceptQualificationRequestResponse smart constructor.

Constructors

AcceptQualificationRequestResponse'

Fields

httpStatus :: Int

The response's http status code.

newAcceptQualificationRequestResponse

Arguments

:: Int
$sel:httpStatus:AcceptQualificationRequestResponse'

-> AcceptQualificationRequestResponse

Create a value of AcceptQualificationRequestResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AcceptQualificationRequestResponse', acceptQualificationRequestResponse_httpStatus - The response's http status code.

Response Lenses

acceptQualificationRequestResponse_httpStatus :: Lens' AcceptQualificationRequestResponse Int

The response's http status code.