Network.AWS.MechanicalTurk.ApproveAssignment

Description

The ApproveAssignment operation approves the results of a completed assignment.

Approving an assignment initiates two payments from the Requester's Amazon.com account

The Worker who submitted the results is paid the reward specified in the HIT.

Amazon Mechanical Turk fees are debited.

If the Requester's account does not have adequate funds for these payments, the call to ApproveAssignment returns an exception, and the approval is not processed. You can include an optional feedback message with the approval, which the Worker can see in the Status section of the web site.

You can also call this operation for assignments that were previous rejected and approve them by explicitly overriding the previous rejection. This only works on rejected assignments that were submitted within the previous 30 days and only if the assignment's related HIT has not been deleted.

Creating a Request

data ApproveAssignment

See: newApproveAssignment smart constructor.

Constructors

ApproveAssignment'

Fields

overrideRejection :: Maybe Bool

A flag indicating that an assignment should be approved even if it was previously rejected. Defaults to False.

requesterFeedback :: Maybe Text

A message for the Worker, which the Worker can see in the Status section of the web site.

assignmentId :: Text

The ID of the assignment. The assignment must correspond to a HIT created by the Requester.

newApproveAssignment

Arguments

:: Text
$sel:assignmentId:ApproveAssignment'

-> ApproveAssignment

Create a value of ApproveAssignment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:overrideRejection:ApproveAssignment', approveAssignment_overrideRejection - A flag indicating that an assignment should be approved even if it was previously rejected. Defaults to False.

$sel:requesterFeedback:ApproveAssignment', approveAssignment_requesterFeedback - A message for the Worker, which the Worker can see in the Status section of the web site.

$sel:assignmentId:ApproveAssignment', approveAssignment_assignmentId - The ID of the assignment. The assignment must correspond to a HIT created by the Requester.

Request Lenses

approveAssignment_overrideRejection :: Lens' ApproveAssignment (Maybe Bool)

A flag indicating that an assignment should be approved even if it was previously rejected. Defaults to False.

approveAssignment_requesterFeedback :: Lens' ApproveAssignment (Maybe Text)

A message for the Worker, which the Worker can see in the Status section of the web site.

approveAssignment_assignmentId :: Lens' ApproveAssignment Text

The ID of the assignment. The assignment must correspond to a HIT created by the Requester.

Destructuring the Response

data ApproveAssignmentResponse

See: newApproveAssignmentResponse smart constructor.

Constructors

ApproveAssignmentResponse'

Fields

httpStatus :: Int

The response's http status code.

newApproveAssignmentResponse

Arguments

:: Int
$sel:httpStatus:ApproveAssignmentResponse'

-> ApproveAssignmentResponse

Create a value of ApproveAssignmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ApproveAssignmentResponse', approveAssignmentResponse_httpStatus - The response's http status code.

Response Lenses

approveAssignmentResponse_httpStatus :: Lens' ApproveAssignmentResponse Int

The response's http status code.