Network.AWS.MechanicalTurk.CreateAdditionalAssignmentsForHIT

Description

The CreateAdditionalAssignmentsForHIT operation increases the maximum number of assignments of an existing HIT.

To extend the maximum number of assignments, specify the number of additional assignments.

HITs created with fewer than 10 assignments cannot be extended to have 10 or more assignments. Attempting to add assignments in a way that brings the total number of assignments for a HIT from fewer than 10 assignments to 10 or more assignments will result in an AWS.MechanicalTurk.InvalidMaximumAssignmentsIncrease exception.

HITs that were created before July 22, 2015 cannot be extended. Attempting to extend HITs that were created before July 22, 2015 will result in an AWS.MechanicalTurk.HITTooOldForExtension exception.

Creating a Request

data CreateAdditionalAssignmentsForHIT

See: newCreateAdditionalAssignmentsForHIT smart constructor.

Constructors

CreateAdditionalAssignmentsForHIT'

Fields

uniqueRequestToken :: Maybe Text

A unique identifier for this request, which allows you to retry the call on error without extending the HIT multiple times. This is useful in cases such as network timeouts where it is unclear whether or not the call succeeded on the server. If the extend HIT already exists in the system from a previous call using the same UniqueRequestToken, subsequent calls will return an error with a message containing the request ID.

hITId :: Text

The ID of the HIT to extend.

numberOfAdditionalAssignments :: Int

The number of additional assignments to request for this HIT.

newCreateAdditionalAssignmentsForHIT

Arguments

:: Text
$sel:hITId:CreateAdditionalAssignmentsForHIT'

-> Int
$sel:numberOfAdditionalAssignments:CreateAdditionalAssignmentsForHIT'

-> CreateAdditionalAssignmentsForHIT

Create a value of CreateAdditionalAssignmentsForHIT with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:uniqueRequestToken:CreateAdditionalAssignmentsForHIT', createAdditionalAssignmentsForHIT_uniqueRequestToken - A unique identifier for this request, which allows you to retry the call on error without extending the HIT multiple times. This is useful in cases such as network timeouts where it is unclear whether or not the call succeeded on the server. If the extend HIT already exists in the system from a previous call using the same UniqueRequestToken, subsequent calls will return an error with a message containing the request ID.

$sel:hITId:CreateAdditionalAssignmentsForHIT', createAdditionalAssignmentsForHIT_hITId - The ID of the HIT to extend.

$sel:numberOfAdditionalAssignments:CreateAdditionalAssignmentsForHIT', createAdditionalAssignmentsForHIT_numberOfAdditionalAssignments - The number of additional assignments to request for this HIT.

Request Lenses

createAdditionalAssignmentsForHIT_uniqueRequestToken :: Lens' CreateAdditionalAssignmentsForHIT (Maybe Text)

A unique identifier for this request, which allows you to retry the call on error without extending the HIT multiple times. This is useful in cases such as network timeouts where it is unclear whether or not the call succeeded on the server. If the extend HIT already exists in the system from a previous call using the same UniqueRequestToken, subsequent calls will return an error with a message containing the request ID.

createAdditionalAssignmentsForHIT_hITId :: Lens' CreateAdditionalAssignmentsForHIT Text

The ID of the HIT to extend.

createAdditionalAssignmentsForHIT_numberOfAdditionalAssignments :: Lens' CreateAdditionalAssignmentsForHIT Int

The number of additional assignments to request for this HIT.

Destructuring the Response

data CreateAdditionalAssignmentsForHITResponse

See: newCreateAdditionalAssignmentsForHITResponse smart constructor.

Constructors

CreateAdditionalAssignmentsForHITResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateAdditionalAssignmentsForHITResponse

Arguments

:: Int
$sel:httpStatus:CreateAdditionalAssignmentsForHITResponse'

-> CreateAdditionalAssignmentsForHITResponse

Create a value of CreateAdditionalAssignmentsForHITResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateAdditionalAssignmentsForHITResponse', createAdditionalAssignmentsForHITResponse_httpStatus - The response's http status code.

Response Lenses

createAdditionalAssignmentsForHITResponse_httpStatus :: Lens' CreateAdditionalAssignmentsForHITResponse Int

The response's http status code.