Network.AWS.MechanicalTurk.CreateHITType

Description

The CreateHITType operation creates a new HIT type. This operation allows you to define a standard set of HIT properties to use when creating HITs. If you register a HIT type with values that match an existing HIT type, the HIT type ID of the existing type will be returned.

Creating a Request

data CreateHITType

See: newCreateHITType smart constructor.

Constructors

CreateHITType'

Fields

autoApprovalDelayInSeconds :: Maybe Integer

The number of seconds after an assignment for the HIT has been submitted, after which the assignment is considered Approved automatically unless the Requester explicitly rejects it.

qualificationRequirements :: Maybe [QualificationRequirement]

Conditions that a Worker's Qualifications must meet in order to accept the HIT. A HIT can have between zero and ten Qualification requirements. All requirements must be met in order for a Worker to accept the HIT. Additionally, other actions can be restricted using the ActionsGuarded field on each QualificationRequirement structure.

keywords :: Maybe Text

One or more words or phrases that describe the HIT, separated by commas. These words are used in searches to find HITs.

assignmentDurationInSeconds :: Integer

The amount of time, in seconds, that a Worker has to complete the HIT after accepting it. If a Worker does not complete the assignment within the specified duration, the assignment is considered abandoned. If the HIT is still active (that is, its lifetime has not elapsed), the assignment becomes available for other users to find and accept.

reward :: Text

The amount of money the Requester will pay a Worker for successfully completing the HIT.

title :: Text

The title of the HIT. A title should be short and descriptive about the kind of task the HIT contains. On the Amazon Mechanical Turk web site, the HIT title appears in search results, and everywhere the HIT is mentioned.

description :: Text

A general description of the HIT. A description includes detailed information about the kind of task the HIT contains. On the Amazon Mechanical Turk web site, the HIT description appears in the expanded view of search results, and in the HIT and assignment screens. A good description gives the user enough information to evaluate the HIT before accepting it.

newCreateHITType

Arguments

:: Integer
$sel:assignmentDurationInSeconds:CreateHITType'

-> Text
$sel:reward:CreateHITType'

-> Text
$sel:title:CreateHITType'

-> Text
$sel:description:CreateHITType'

-> CreateHITType

Create a value of CreateHITType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:autoApprovalDelayInSeconds:CreateHITType', createHITType_autoApprovalDelayInSeconds - The number of seconds after an assignment for the HIT has been submitted, after which the assignment is considered Approved automatically unless the Requester explicitly rejects it.

$sel:qualificationRequirements:CreateHITType', createHITType_qualificationRequirements - Conditions that a Worker's Qualifications must meet in order to accept the HIT. A HIT can have between zero and ten Qualification requirements. All requirements must be met in order for a Worker to accept the HIT. Additionally, other actions can be restricted using the ActionsGuarded field on each QualificationRequirement structure.

$sel:keywords:CreateHITType', createHITType_keywords - One or more words or phrases that describe the HIT, separated by commas. These words are used in searches to find HITs.

$sel:assignmentDurationInSeconds:CreateHITType', createHITType_assignmentDurationInSeconds - The amount of time, in seconds, that a Worker has to complete the HIT after accepting it. If a Worker does not complete the assignment within the specified duration, the assignment is considered abandoned. If the HIT is still active (that is, its lifetime has not elapsed), the assignment becomes available for other users to find and accept.

$sel:reward:CreateHITType', createHITType_reward - The amount of money the Requester will pay a Worker for successfully completing the HIT.

$sel:title:CreateHITType', createHITType_title - The title of the HIT. A title should be short and descriptive about the kind of task the HIT contains. On the Amazon Mechanical Turk web site, the HIT title appears in search results, and everywhere the HIT is mentioned.

$sel:description:CreateHITType', createHITType_description - A general description of the HIT. A description includes detailed information about the kind of task the HIT contains. On the Amazon Mechanical Turk web site, the HIT description appears in the expanded view of search results, and in the HIT and assignment screens. A good description gives the user enough information to evaluate the HIT before accepting it.

Request Lenses

createHITType_autoApprovalDelayInSeconds :: Lens' CreateHITType (Maybe Integer)

The number of seconds after an assignment for the HIT has been submitted, after which the assignment is considered Approved automatically unless the Requester explicitly rejects it.

createHITType_qualificationRequirements :: Lens' CreateHITType (Maybe [QualificationRequirement])

Conditions that a Worker's Qualifications must meet in order to accept the HIT. A HIT can have between zero and ten Qualification requirements. All requirements must be met in order for a Worker to accept the HIT. Additionally, other actions can be restricted using the ActionsGuarded field on each QualificationRequirement structure.

createHITType_keywords :: Lens' CreateHITType (Maybe Text)

One or more words or phrases that describe the HIT, separated by commas. These words are used in searches to find HITs.

createHITType_assignmentDurationInSeconds :: Lens' CreateHITType Integer

The amount of time, in seconds, that a Worker has to complete the HIT after accepting it. If a Worker does not complete the assignment within the specified duration, the assignment is considered abandoned. If the HIT is still active (that is, its lifetime has not elapsed), the assignment becomes available for other users to find and accept.

createHITType_reward :: Lens' CreateHITType Text

The amount of money the Requester will pay a Worker for successfully completing the HIT.

createHITType_title :: Lens' CreateHITType Text

The title of the HIT. A title should be short and descriptive about the kind of task the HIT contains. On the Amazon Mechanical Turk web site, the HIT title appears in search results, and everywhere the HIT is mentioned.

createHITType_description :: Lens' CreateHITType Text

A general description of the HIT. A description includes detailed information about the kind of task the HIT contains. On the Amazon Mechanical Turk web site, the HIT description appears in the expanded view of search results, and in the HIT and assignment screens. A good description gives the user enough information to evaluate the HIT before accepting it.

Destructuring the Response

data CreateHITTypeResponse

See: newCreateHITTypeResponse smart constructor.

Constructors

CreateHITTypeResponse'

Fields

hITTypeId :: Maybe Text

The ID of the newly registered HIT type.

httpStatus :: Int

The response's http status code.

newCreateHITTypeResponse

Arguments

:: Int
$sel:httpStatus:CreateHITTypeResponse'

-> CreateHITTypeResponse

Create a value of CreateHITTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hITTypeId:CreateHITTypeResponse', createHITTypeResponse_hITTypeId - The ID of the newly registered HIT type.

$sel:httpStatus:CreateHITTypeResponse', createHITTypeResponse_httpStatus - The response's http status code.

Response Lenses

createHITTypeResponse_hITTypeId :: Lens' CreateHITTypeResponse (Maybe Text)

The ID of the newly registered HIT type.

createHITTypeResponse_httpStatus :: Lens' CreateHITTypeResponse Int

The response's http status code.