Network.AWS.MechanicalTurk.DeleteWorkerBlock

Description

The DeleteWorkerBlock operation allows you to reinstate a blocked Worker to work on your HITs. This operation reverses the effects of the CreateWorkerBlock operation. You need the Worker ID to use this operation. If the Worker ID is missing or invalid, this operation fails and returns the message “WorkerId is invalid.” If the specified Worker is not blocked, this operation returns successfully.

Creating a Request

data DeleteWorkerBlock

See: newDeleteWorkerBlock smart constructor.

Constructors

DeleteWorkerBlock'

Fields

reason :: Maybe Text

A message that explains the reason for unblocking the Worker. The Worker does not see this message.

workerId :: Text

The ID of the Worker to unblock.

newDeleteWorkerBlock

Arguments

:: Text
$sel:workerId:DeleteWorkerBlock'

-> DeleteWorkerBlock

Create a value of DeleteWorkerBlock with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reason:DeleteWorkerBlock', deleteWorkerBlock_reason - A message that explains the reason for unblocking the Worker. The Worker does not see this message.

$sel:workerId:DeleteWorkerBlock', deleteWorkerBlock_workerId - The ID of the Worker to unblock.

Request Lenses

deleteWorkerBlock_reason :: Lens' DeleteWorkerBlock (Maybe Text)

A message that explains the reason for unblocking the Worker. The Worker does not see this message.

deleteWorkerBlock_workerId :: Lens' DeleteWorkerBlock Text

The ID of the Worker to unblock.

Destructuring the Response

data DeleteWorkerBlockResponse

See: newDeleteWorkerBlockResponse smart constructor.

Constructors

DeleteWorkerBlockResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteWorkerBlockResponse

Arguments

:: Int
$sel:httpStatus:DeleteWorkerBlockResponse'

-> DeleteWorkerBlockResponse

Create a value of DeleteWorkerBlockResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteWorkerBlockResponse', deleteWorkerBlockResponse_httpStatus - The response's http status code.

Response Lenses

deleteWorkerBlockResponse_httpStatus :: Lens' DeleteWorkerBlockResponse Int

The response's http status code.