Network.AWS.MechanicalTurk.GetFileUploadURL

Description

The GetFileUploadURL operation generates and returns a temporary URL. You use the temporary URL to retrieve a file uploaded by a Worker as an answer to a FileUploadAnswer question for a HIT. The temporary URL is generated the instant the GetFileUploadURL operation is called, and is valid for 60 seconds. You can get a temporary file upload URL any time until the HIT is disposed. After the HIT is disposed, any uploaded files are deleted, and cannot be retrieved. Pending Deprecation on December 12, 2017. The Answer Specification structure will no longer support the FileUploadAnswer element to be used for the QuestionForm data structure. Instead, we recommend that Requesters who want to create HITs asking Workers to upload files to use Amazon S3.

Creating a Request

data GetFileUploadURL

See: newGetFileUploadURL smart constructor.

Constructors

GetFileUploadURL'

Fields

assignmentId :: Text

The ID of the assignment that contains the question with a FileUploadAnswer.

questionIdentifier :: Text

The identifier of the question with a FileUploadAnswer, as specified in the QuestionForm of the HIT.

newGetFileUploadURL

Arguments

:: Text
$sel:assignmentId:GetFileUploadURL'

-> Text
$sel:questionIdentifier:GetFileUploadURL'

-> GetFileUploadURL

Create a value of GetFileUploadURL with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assignmentId:GetFileUploadURL', getFileUploadURL_assignmentId - The ID of the assignment that contains the question with a FileUploadAnswer.

$sel:questionIdentifier:GetFileUploadURL', getFileUploadURL_questionIdentifier - The identifier of the question with a FileUploadAnswer, as specified in the QuestionForm of the HIT.

Request Lenses

getFileUploadURL_assignmentId :: Lens' GetFileUploadURL Text

The ID of the assignment that contains the question with a FileUploadAnswer.

getFileUploadURL_questionIdentifier :: Lens' GetFileUploadURL Text

The identifier of the question with a FileUploadAnswer, as specified in the QuestionForm of the HIT.

Destructuring the Response

data GetFileUploadURLResponse

See: newGetFileUploadURLResponse smart constructor.

Constructors

GetFileUploadURLResponse'

Fields

fileUploadURL :: Maybe Text

A temporary URL for the file that the Worker uploaded for the answer.

httpStatus :: Int

The response's http status code.

newGetFileUploadURLResponse

Arguments

:: Int
$sel:httpStatus:GetFileUploadURLResponse'

-> GetFileUploadURLResponse

Create a value of GetFileUploadURLResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fileUploadURL:GetFileUploadURLResponse', getFileUploadURLResponse_fileUploadURL - A temporary URL for the file that the Worker uploaded for the answer.

$sel:httpStatus:GetFileUploadURLResponse', getFileUploadURLResponse_httpStatus - The response's http status code.

Response Lenses

getFileUploadURLResponse_fileUploadURL :: Lens' GetFileUploadURLResponse (Maybe Text)

A temporary URL for the file that the Worker uploaded for the answer.

getFileUploadURLResponse_httpStatus :: Lens' GetFileUploadURLResponse Int

The response's http status code.