Network.AWS.MechanicalTurk.GetQualificationScore

Description

The GetQualificationScore operation returns the value of a Worker's Qualification for a given Qualification type.

To get a Worker's Qualification, you must know the Worker's ID. The Worker's ID is included in the assignment data returned by the ListAssignmentsForHIT operation.

Only the owner of a Qualification type can query the value of a Worker's Qualification of that type.

Creating a Request

data GetQualificationScore

See: newGetQualificationScore smart constructor.

Constructors

GetQualificationScore'

Fields

qualificationTypeId :: Text

The ID of the QualificationType.

workerId :: Text

The ID of the Worker whose Qualification is being updated.

newGetQualificationScore

Arguments

:: Text
$sel:qualificationTypeId:GetQualificationScore'

-> Text
$sel:workerId:GetQualificationScore'

-> GetQualificationScore

Create a value of GetQualificationScore with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:qualificationTypeId:GetQualificationScore', getQualificationScore_qualificationTypeId - The ID of the QualificationType.

$sel:workerId:GetQualificationScore', getQualificationScore_workerId - The ID of the Worker whose Qualification is being updated.

Request Lenses

getQualificationScore_qualificationTypeId :: Lens' GetQualificationScore Text

The ID of the QualificationType.

getQualificationScore_workerId :: Lens' GetQualificationScore Text

The ID of the Worker whose Qualification is being updated.

Destructuring the Response

data GetQualificationScoreResponse

See: newGetQualificationScoreResponse smart constructor.

Constructors

GetQualificationScoreResponse'

Fields

qualification :: Maybe Qualification

The Qualification data structure of the Qualification assigned to a user, including the Qualification type and the value (score).

httpStatus :: Int

The response's http status code.

newGetQualificationScoreResponse

Arguments

:: Int
$sel:httpStatus:GetQualificationScoreResponse'

-> GetQualificationScoreResponse

Create a value of GetQualificationScoreResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:qualification:GetQualificationScoreResponse', getQualificationScoreResponse_qualification - The Qualification data structure of the Qualification assigned to a user, including the Qualification type and the value (score).

$sel:httpStatus:GetQualificationScoreResponse', getQualificationScoreResponse_httpStatus - The response's http status code.

Response Lenses

getQualificationScoreResponse_qualification :: Lens' GetQualificationScoreResponse (Maybe Qualification)

The Qualification data structure of the Qualification assigned to a user, including the Qualification type and the value (score).

getQualificationScoreResponse_httpStatus :: Lens' GetQualificationScoreResponse Int

The response's http status code.