Network.AWS.MechanicalTurk.ListAssignmentsForHIT

Description

The ListAssignmentsForHIT operation retrieves completed assignments for a HIT. You can use this operation to retrieve the results for a HIT.

You can get assignments for a HIT at any time, even if the HIT is not yet Reviewable. If a HIT requested multiple assignments, and has received some results but has not yet become Reviewable, you can still retrieve the partial results with this operation.

Use the AssignmentStatus parameter to control which set of assignments for a HIT are returned. The ListAssignmentsForHIT operation can return submitted assignments awaiting approval, or it can return assignments that have already been approved or rejected. You can set AssignmentStatus=Approved,Rejected to get assignments that have already been approved and rejected together in one result set.

Only the Requester who created the HIT can retrieve the assignments for that HIT.

Results are sorted and divided into numbered pages and the operation returns a single page of results. You can use the parameters of the operation to control sorting and pagination.

This operation returns paginated results.

Creating a Request

data ListAssignmentsForHIT

See: newListAssignmentsForHIT smart constructor.

Constructors

ListAssignmentsForHIT'

Fields

assignmentStatuses :: Maybe [AssignmentStatus]

The status of the assignments to return: Submitted | Approved | Rejected

nextToken :: Maybe Text

Pagination token

maxResults :: Maybe Natural

hITId :: Text

The ID of the HIT.

newListAssignmentsForHIT

Arguments

:: Text
$sel:hITId:ListAssignmentsForHIT'

-> ListAssignmentsForHIT

Create a value of ListAssignmentsForHIT with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assignmentStatuses:ListAssignmentsForHIT', listAssignmentsForHIT_assignmentStatuses - The status of the assignments to return: Submitted | Approved | Rejected

$sel:nextToken:ListAssignmentsForHIT', listAssignmentsForHIT_nextToken - Pagination token

$sel:maxResults:ListAssignmentsForHIT', listAssignmentsForHIT_maxResults - Undocumented member.

$sel:hITId:ListAssignmentsForHIT', listAssignmentsForHIT_hITId - The ID of the HIT.

Request Lenses

listAssignmentsForHIT_assignmentStatuses :: Lens' ListAssignmentsForHIT (Maybe [AssignmentStatus])

The status of the assignments to return: Submitted | Approved | Rejected

listAssignmentsForHIT_nextToken :: Lens' ListAssignmentsForHIT (Maybe Text)

Pagination token

listAssignmentsForHIT_maxResults :: Lens' ListAssignmentsForHIT (Maybe Natural)

Undocumented member.

listAssignmentsForHIT_hITId :: Lens' ListAssignmentsForHIT Text

The ID of the HIT.

Destructuring the Response

data ListAssignmentsForHITResponse

See: newListAssignmentsForHITResponse smart constructor.

Constructors

ListAssignmentsForHITResponse'

Fields

nextToken :: Maybe Text

numResults :: Maybe Int

The number of assignments on the page in the filtered results list, equivalent to the number of assignments returned by this call.

assignments :: Maybe [Assignment]

The collection of Assignment data structures returned by this call.

httpStatus :: Int

The response's http status code.

newListAssignmentsForHITResponse

Arguments

:: Int
$sel:httpStatus:ListAssignmentsForHITResponse'

-> ListAssignmentsForHITResponse

Create a value of ListAssignmentsForHITResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAssignmentsForHIT', listAssignmentsForHITResponse_nextToken - Undocumented member.

$sel:numResults:ListAssignmentsForHITResponse', listAssignmentsForHITResponse_numResults - The number of assignments on the page in the filtered results list, equivalent to the number of assignments returned by this call.

$sel:assignments:ListAssignmentsForHITResponse', listAssignmentsForHITResponse_assignments - The collection of Assignment data structures returned by this call.

$sel:httpStatus:ListAssignmentsForHITResponse', listAssignmentsForHITResponse_httpStatus - The response's http status code.

Response Lenses

listAssignmentsForHITResponse_nextToken :: Lens' ListAssignmentsForHITResponse (Maybe Text)

Undocumented member.

listAssignmentsForHITResponse_numResults :: Lens' ListAssignmentsForHITResponse (Maybe Int)

The number of assignments on the page in the filtered results list, equivalent to the number of assignments returned by this call.

listAssignmentsForHITResponse_assignments :: Lens' ListAssignmentsForHITResponse (Maybe [Assignment])

The collection of Assignment data structures returned by this call.

listAssignmentsForHITResponse_httpStatus :: Lens' ListAssignmentsForHITResponse Int

The response's http status code.