Network.AWS.MechanicalTurk.ListBonusPayments

Description

The ListBonusPayments operation retrieves the amounts of bonuses you have paid to Workers for a given HIT or assignment.

This operation returns paginated results.

Creating a Request

data ListBonusPayments

See: newListBonusPayments smart constructor.

Constructors

ListBonusPayments'

Fields

nextToken :: Maybe Text

Pagination token

assignmentId :: Maybe Text

The ID of the assignment associated with the bonus payments to retrieve. If specified, only bonus payments for the given assignment are returned. Either the HITId parameter or the AssignmentId parameter must be specified

maxResults :: Maybe Natural

hITId :: Maybe Text

The ID of the HIT associated with the bonus payments to retrieve. If not specified, all bonus payments for all assignments for the given HIT are returned. Either the HITId parameter or the AssignmentId parameter must be specified

newListBonusPayments :: ListBonusPayments

Create a value of ListBonusPayments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListBonusPayments', listBonusPayments_nextToken - Pagination token

$sel:assignmentId:ListBonusPayments', listBonusPayments_assignmentId - The ID of the assignment associated with the bonus payments to retrieve. If specified, only bonus payments for the given assignment are returned. Either the HITId parameter or the AssignmentId parameter must be specified

$sel:maxResults:ListBonusPayments', listBonusPayments_maxResults - Undocumented member.

$sel:hITId:ListBonusPayments', listBonusPayments_hITId - The ID of the HIT associated with the bonus payments to retrieve. If not specified, all bonus payments for all assignments for the given HIT are returned. Either the HITId parameter or the AssignmentId parameter must be specified

Request Lenses

listBonusPayments_nextToken :: Lens' ListBonusPayments (Maybe Text)

Pagination token

listBonusPayments_assignmentId :: Lens' ListBonusPayments (Maybe Text)

The ID of the assignment associated with the bonus payments to retrieve. If specified, only bonus payments for the given assignment are returned. Either the HITId parameter or the AssignmentId parameter must be specified

listBonusPayments_maxResults :: Lens' ListBonusPayments (Maybe Natural)

Undocumented member.

listBonusPayments_hITId :: Lens' ListBonusPayments (Maybe Text)

The ID of the HIT associated with the bonus payments to retrieve. If not specified, all bonus payments for all assignments for the given HIT are returned. Either the HITId parameter or the AssignmentId parameter must be specified

Destructuring the Response

data ListBonusPaymentsResponse

See: newListBonusPaymentsResponse smart constructor.

Constructors

ListBonusPaymentsResponse'

Fields

nextToken :: Maybe Text

numResults :: Maybe Int

The number of bonus payments on this page in the filtered results list, equivalent to the number of bonus payments being returned by this call.

bonusPayments :: Maybe [BonusPayment]

A successful request to the ListBonusPayments operation returns a list of BonusPayment objects.

httpStatus :: Int

The response's http status code.

newListBonusPaymentsResponse

Arguments

:: Int
$sel:httpStatus:ListBonusPaymentsResponse'

-> ListBonusPaymentsResponse

Create a value of ListBonusPaymentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListBonusPayments', listBonusPaymentsResponse_nextToken - Undocumented member.

$sel:numResults:ListBonusPaymentsResponse', listBonusPaymentsResponse_numResults - The number of bonus payments on this page in the filtered results list, equivalent to the number of bonus payments being returned by this call.

$sel:bonusPayments:ListBonusPaymentsResponse', listBonusPaymentsResponse_bonusPayments - A successful request to the ListBonusPayments operation returns a list of BonusPayment objects.

$sel:httpStatus:ListBonusPaymentsResponse', listBonusPaymentsResponse_httpStatus - The response's http status code.

Response Lenses

listBonusPaymentsResponse_nextToken :: Lens' ListBonusPaymentsResponse (Maybe Text)

Undocumented member.

listBonusPaymentsResponse_numResults :: Lens' ListBonusPaymentsResponse (Maybe Int)

The number of bonus payments on this page in the filtered results list, equivalent to the number of bonus payments being returned by this call.

listBonusPaymentsResponse_bonusPayments :: Lens' ListBonusPaymentsResponse (Maybe [BonusPayment])

A successful request to the ListBonusPayments operation returns a list of BonusPayment objects.

listBonusPaymentsResponse_httpStatus :: Lens' ListBonusPaymentsResponse Int

The response's http status code.