Network.AWS.MechanicalTurk.ListReviewPolicyResultsForHIT

Description

The ListReviewPolicyResultsForHIT operation retrieves the computed results and the actions taken in the course of executing your Review Policies for a given HIT. For information about how to specify Review Policies when you call CreateHIT, see Review Policies. The ListReviewPolicyResultsForHIT operation can return results for both Assignment-level and HIT-level review results.

Creating a Request

data ListReviewPolicyResultsForHIT

See: newListReviewPolicyResultsForHIT smart constructor.

Constructors

ListReviewPolicyResultsForHIT'

Fields

nextToken :: Maybe Text

Pagination token

maxResults :: Maybe Natural

Limit the number of results returned.

retrieveResults :: Maybe Bool

Specify if the operation should retrieve a list of the results computed by the Review Policies.

retrieveActions :: Maybe Bool

Specify if the operation should retrieve a list of the actions taken executing the Review Policies and their outcomes.

policyLevels :: Maybe [ReviewPolicyLevel]

The Policy Level(s) to retrieve review results for - HIT or Assignment. If omitted, the default behavior is to retrieve all data for both policy levels. For a list of all the described policies, see Review Policies.

hITId :: Text

The unique identifier of the HIT to retrieve review results for.

newListReviewPolicyResultsForHIT

Arguments

:: Text
$sel:hITId:ListReviewPolicyResultsForHIT'

-> ListReviewPolicyResultsForHIT

Create a value of ListReviewPolicyResultsForHIT with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListReviewPolicyResultsForHIT', listReviewPolicyResultsForHIT_nextToken - Pagination token

$sel:maxResults:ListReviewPolicyResultsForHIT', listReviewPolicyResultsForHIT_maxResults - Limit the number of results returned.

$sel:retrieveResults:ListReviewPolicyResultsForHIT', listReviewPolicyResultsForHIT_retrieveResults - Specify if the operation should retrieve a list of the results computed by the Review Policies.

$sel:retrieveActions:ListReviewPolicyResultsForHIT', listReviewPolicyResultsForHIT_retrieveActions - Specify if the operation should retrieve a list of the actions taken executing the Review Policies and their outcomes.

$sel:policyLevels:ListReviewPolicyResultsForHIT', listReviewPolicyResultsForHIT_policyLevels - The Policy Level(s) to retrieve review results for - HIT or Assignment. If omitted, the default behavior is to retrieve all data for both policy levels. For a list of all the described policies, see Review Policies.

$sel:hITId:ListReviewPolicyResultsForHIT', listReviewPolicyResultsForHIT_hITId - The unique identifier of the HIT to retrieve review results for.

Request Lenses

listReviewPolicyResultsForHIT_nextToken :: Lens' ListReviewPolicyResultsForHIT (Maybe Text)

Pagination token

listReviewPolicyResultsForHIT_maxResults :: Lens' ListReviewPolicyResultsForHIT (Maybe Natural)

Limit the number of results returned.

listReviewPolicyResultsForHIT_retrieveResults :: Lens' ListReviewPolicyResultsForHIT (Maybe Bool)

Specify if the operation should retrieve a list of the results computed by the Review Policies.

listReviewPolicyResultsForHIT_retrieveActions :: Lens' ListReviewPolicyResultsForHIT (Maybe Bool)

Specify if the operation should retrieve a list of the actions taken executing the Review Policies and their outcomes.

listReviewPolicyResultsForHIT_policyLevels :: Lens' ListReviewPolicyResultsForHIT (Maybe [ReviewPolicyLevel])

The Policy Level(s) to retrieve review results for - HIT or Assignment. If omitted, the default behavior is to retrieve all data for both policy levels. For a list of all the described policies, see Review Policies.

listReviewPolicyResultsForHIT_hITId :: Lens' ListReviewPolicyResultsForHIT Text

The unique identifier of the HIT to retrieve review results for.

Destructuring the Response

data ListReviewPolicyResultsForHITResponse

See: newListReviewPolicyResultsForHITResponse smart constructor.

Constructors

ListReviewPolicyResultsForHITResponse'

Fields

nextToken :: Maybe Text

hITId :: Maybe Text

The HITId of the HIT for which results have been returned.

hITReviewPolicy :: Maybe ReviewPolicy

The name of the HIT-level Review Policy. This contains only the PolicyName element.

assignmentReviewReport :: Maybe ReviewReport

Contains both ReviewResult and ReviewAction elements for an Assignment.

hITReviewReport :: Maybe ReviewReport

Contains both ReviewResult and ReviewAction elements for a particular HIT.

assignmentReviewPolicy :: Maybe ReviewPolicy

The name of the Assignment-level Review Policy. This contains only the PolicyName element.

httpStatus :: Int

The response's http status code.

newListReviewPolicyResultsForHITResponse

Arguments

:: Int
$sel:httpStatus:ListReviewPolicyResultsForHITResponse'

-> ListReviewPolicyResultsForHITResponse

Create a value of ListReviewPolicyResultsForHITResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListReviewPolicyResultsForHIT', listReviewPolicyResultsForHITResponse_nextToken - Undocumented member.

$sel:hITId:ListReviewPolicyResultsForHIT', listReviewPolicyResultsForHITResponse_hITId - The HITId of the HIT for which results have been returned.

$sel:hITReviewPolicy:ListReviewPolicyResultsForHITResponse', listReviewPolicyResultsForHITResponse_hITReviewPolicy - The name of the HIT-level Review Policy. This contains only the PolicyName element.

$sel:assignmentReviewReport:ListReviewPolicyResultsForHITResponse', listReviewPolicyResultsForHITResponse_assignmentReviewReport - Contains both ReviewResult and ReviewAction elements for an Assignment.

$sel:hITReviewReport:ListReviewPolicyResultsForHITResponse', listReviewPolicyResultsForHITResponse_hITReviewReport - Contains both ReviewResult and ReviewAction elements for a particular HIT.

$sel:assignmentReviewPolicy:ListReviewPolicyResultsForHITResponse', listReviewPolicyResultsForHITResponse_assignmentReviewPolicy - The name of the Assignment-level Review Policy. This contains only the PolicyName element.

$sel:httpStatus:ListReviewPolicyResultsForHITResponse', listReviewPolicyResultsForHITResponse_httpStatus - The response's http status code.

Response Lenses

listReviewPolicyResultsForHITResponse_nextToken :: Lens' ListReviewPolicyResultsForHITResponse (Maybe Text)

Undocumented member.

listReviewPolicyResultsForHITResponse_hITId :: Lens' ListReviewPolicyResultsForHITResponse (Maybe Text)

The HITId of the HIT for which results have been returned.

listReviewPolicyResultsForHITResponse_hITReviewPolicy :: Lens' ListReviewPolicyResultsForHITResponse (Maybe ReviewPolicy)

The name of the HIT-level Review Policy. This contains only the PolicyName element.

listReviewPolicyResultsForHITResponse_assignmentReviewReport :: Lens' ListReviewPolicyResultsForHITResponse (Maybe ReviewReport)

Contains both ReviewResult and ReviewAction elements for an Assignment.

listReviewPolicyResultsForHITResponse_hITReviewReport :: Lens' ListReviewPolicyResultsForHITResponse (Maybe ReviewReport)

Contains both ReviewResult and ReviewAction elements for a particular HIT.

listReviewPolicyResultsForHITResponse_assignmentReviewPolicy :: Lens' ListReviewPolicyResultsForHITResponse (Maybe ReviewPolicy)

The name of the Assignment-level Review Policy. This contains only the PolicyName element.

listReviewPolicyResultsForHITResponse_httpStatus :: Lens' ListReviewPolicyResultsForHITResponse Int

The response's http status code.