Network.AWS.MechanicalTurk.ListWorkerBlocks

Description

The ListWorkersBlocks operation retrieves a list of Workers who are blocked from working on your HITs.

This operation returns paginated results.

Creating a Request

data ListWorkerBlocks

See: newListWorkerBlocks smart constructor.

Constructors

ListWorkerBlocks'

Fields

nextToken :: Maybe Text

Pagination token

maxResults :: Maybe Natural

newListWorkerBlocks :: ListWorkerBlocks

Create a value of ListWorkerBlocks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListWorkerBlocks', listWorkerBlocks_nextToken - Pagination token

$sel:maxResults:ListWorkerBlocks', listWorkerBlocks_maxResults - Undocumented member.

Request Lenses

listWorkerBlocks_nextToken :: Lens' ListWorkerBlocks (Maybe Text)

Pagination token

listWorkerBlocks_maxResults :: Lens' ListWorkerBlocks (Maybe Natural)

Undocumented member.

Destructuring the Response

data ListWorkerBlocksResponse

See: newListWorkerBlocksResponse smart constructor.

Constructors

ListWorkerBlocksResponse'

Fields

nextToken :: Maybe Text

workerBlocks :: Maybe [WorkerBlock]

The list of WorkerBlocks, containing the collection of Worker IDs and reasons for blocking.

numResults :: Maybe Int

The number of assignments on the page in the filtered results list, equivalent to the number of assignments returned by this call.

httpStatus :: Int

The response's http status code.

newListWorkerBlocksResponse

Arguments

:: Int
$sel:httpStatus:ListWorkerBlocksResponse'

-> ListWorkerBlocksResponse

Create a value of ListWorkerBlocksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListWorkerBlocks', listWorkerBlocksResponse_nextToken - Undocumented member.

$sel:workerBlocks:ListWorkerBlocksResponse', listWorkerBlocksResponse_workerBlocks - The list of WorkerBlocks, containing the collection of Worker IDs and reasons for blocking.

$sel:numResults:ListWorkerBlocksResponse', listWorkerBlocksResponse_numResults - The number of assignments on the page in the filtered results list, equivalent to the number of assignments returned by this call.

$sel:httpStatus:ListWorkerBlocksResponse', listWorkerBlocksResponse_httpStatus - The response's http status code.

Response Lenses

listWorkerBlocksResponse_nextToken :: Lens' ListWorkerBlocksResponse (Maybe Text)

Undocumented member.

listWorkerBlocksResponse_workerBlocks :: Lens' ListWorkerBlocksResponse (Maybe [WorkerBlock])

The list of WorkerBlocks, containing the collection of Worker IDs and reasons for blocking.

listWorkerBlocksResponse_numResults :: Lens' ListWorkerBlocksResponse (Maybe Int)

The number of assignments on the page in the filtered results list, equivalent to the number of assignments returned by this call.

listWorkerBlocksResponse_httpStatus :: Lens' ListWorkerBlocksResponse Int

The response's http status code.