Network.AWS.MechanicalTurk.ListWorkersWithQualificationType

Description

The ListWorkersWithQualificationType operation returns all of the Workers that have been associated with a given Qualification type.

This operation returns paginated results.

Creating a Request

data ListWorkersWithQualificationType

See: newListWorkersWithQualificationType smart constructor.

Constructors

ListWorkersWithQualificationType'

Fields

status :: Maybe QualificationStatus

The status of the Qualifications to return. Can be Granted | Revoked.

nextToken :: Maybe Text

Pagination Token

maxResults :: Maybe Natural

Limit the number of results returned.

qualificationTypeId :: Text

The ID of the Qualification type of the Qualifications to return.

newListWorkersWithQualificationType

Arguments

:: Text
$sel:qualificationTypeId:ListWorkersWithQualificationType'

-> ListWorkersWithQualificationType

Create a value of ListWorkersWithQualificationType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListWorkersWithQualificationType', listWorkersWithQualificationType_status - The status of the Qualifications to return. Can be Granted | Revoked.

$sel:nextToken:ListWorkersWithQualificationType', listWorkersWithQualificationType_nextToken - Pagination Token

$sel:maxResults:ListWorkersWithQualificationType', listWorkersWithQualificationType_maxResults - Limit the number of results returned.

$sel:qualificationTypeId:ListWorkersWithQualificationType', listWorkersWithQualificationType_qualificationTypeId - The ID of the Qualification type of the Qualifications to return.

Request Lenses

listWorkersWithQualificationType_status :: Lens' ListWorkersWithQualificationType (Maybe QualificationStatus)

The status of the Qualifications to return. Can be Granted | Revoked.

listWorkersWithQualificationType_nextToken :: Lens' ListWorkersWithQualificationType (Maybe Text)

Pagination Token

listWorkersWithQualificationType_maxResults :: Lens' ListWorkersWithQualificationType (Maybe Natural)

Limit the number of results returned.

listWorkersWithQualificationType_qualificationTypeId :: Lens' ListWorkersWithQualificationType Text

The ID of the Qualification type of the Qualifications to return.

Destructuring the Response

data ListWorkersWithQualificationTypeResponse

See: newListWorkersWithQualificationTypeResponse smart constructor.

Constructors

ListWorkersWithQualificationTypeResponse'

Fields

nextToken :: Maybe Text

numResults :: Maybe Int

The number of Qualifications on this page in the filtered results list, equivalent to the number of Qualifications being returned by this call.

qualifications :: Maybe [Qualification]

The list of Qualification elements returned by this call.

httpStatus :: Int

The response's http status code.

newListWorkersWithQualificationTypeResponse

Arguments

:: Int
$sel:httpStatus:ListWorkersWithQualificationTypeResponse'

-> ListWorkersWithQualificationTypeResponse

Create a value of ListWorkersWithQualificationTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListWorkersWithQualificationType', listWorkersWithQualificationTypeResponse_nextToken - Undocumented member.

$sel:numResults:ListWorkersWithQualificationTypeResponse', listWorkersWithQualificationTypeResponse_numResults - The number of Qualifications on this page in the filtered results list, equivalent to the number of Qualifications being returned by this call.

$sel:qualifications:ListWorkersWithQualificationTypeResponse', listWorkersWithQualificationTypeResponse_qualifications - The list of Qualification elements returned by this call.

$sel:httpStatus:ListWorkersWithQualificationTypeResponse', listWorkersWithQualificationTypeResponse_httpStatus - The response's http status code.

Response Lenses

listWorkersWithQualificationTypeResponse_nextToken :: Lens' ListWorkersWithQualificationTypeResponse (Maybe Text)

Undocumented member.

listWorkersWithQualificationTypeResponse_numResults :: Lens' ListWorkersWithQualificationTypeResponse (Maybe Int)

The number of Qualifications on this page in the filtered results list, equivalent to the number of Qualifications being returned by this call.

listWorkersWithQualificationTypeResponse_qualifications :: Lens' ListWorkersWithQualificationTypeResponse (Maybe [Qualification])

The list of Qualification elements returned by this call.

listWorkersWithQualificationTypeResponse_httpStatus :: Lens' ListWorkersWithQualificationTypeResponse Int

The response's http status code.