Network.AWS.MechanicalTurk.Types.Assignment

Description

Documentation

data Assignment

The Assignment data structure represents a single assignment of a HIT to a Worker. The assignment tracks the Worker's efforts to complete the HIT, and contains the results for later retrieval.

See: newAssignment smart constructor.

Constructors

Assignment'

Fields

requesterFeedback :: Maybe Text

The feedback string included with the call to the ApproveAssignment operation or the RejectAssignment operation, if the Requester approved or rejected the assignment and specified feedback.

assignmentId :: Maybe Text

A unique identifier for the assignment.

hITId :: Maybe Text

The ID of the HIT.

autoApprovalTime :: Maybe POSIX

If results have been submitted, AutoApprovalTime is the date and time the results of the assignment results are considered Approved automatically if they have not already been explicitly approved or rejected by the Requester. This value is derived from the auto-approval delay specified by the Requester in the HIT. This value is omitted from the assignment if the Worker has not yet submitted results.

rejectionTime :: Maybe POSIX

If the Worker has submitted results and the Requester has rejected the results, RejectionTime is the date and time the Requester rejected the results.

assignmentStatus :: Maybe AssignmentStatus

The status of the assignment.

answer :: Maybe Text

The Worker's answers submitted for the HIT contained in a QuestionFormAnswers document, if the Worker provides an answer. If the Worker does not provide any answers, Answer may contain a QuestionFormAnswers document, or Answer may be empty.

submitTime :: Maybe POSIX

If the Worker has submitted results, SubmitTime is the date and time the assignment was submitted. This value is omitted from the assignment if the Worker has not yet submitted results.

workerId :: Maybe Text

The ID of the Worker who accepted the HIT.

acceptTime :: Maybe POSIX

The date and time the Worker accepted the assignment.

deadline :: Maybe POSIX

The date and time of the deadline for the assignment. This value is derived from the deadline specification for the HIT and the date and time the Worker accepted the HIT.

approvalTime :: Maybe POSIX

If the Worker has submitted results and the Requester has approved the results, ApprovalTime is the date and time the Requester approved the results. This value is omitted from the assignment if the Requester has not yet approved the results.

newAssignment :: Assignment

Create a value of Assignment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requesterFeedback:Assignment', assignment_requesterFeedback - The feedback string included with the call to the ApproveAssignment operation or the RejectAssignment operation, if the Requester approved or rejected the assignment and specified feedback.

$sel:assignmentId:Assignment', assignment_assignmentId - A unique identifier for the assignment.

$sel:hITId:Assignment', assignment_hITId - The ID of the HIT.

$sel:autoApprovalTime:Assignment', assignment_autoApprovalTime - If results have been submitted, AutoApprovalTime is the date and time the results of the assignment results are considered Approved automatically if they have not already been explicitly approved or rejected by the Requester. This value is derived from the auto-approval delay specified by the Requester in the HIT. This value is omitted from the assignment if the Worker has not yet submitted results.

$sel:rejectionTime:Assignment', assignment_rejectionTime - If the Worker has submitted results and the Requester has rejected the results, RejectionTime is the date and time the Requester rejected the results.

$sel:assignmentStatus:Assignment', assignment_assignmentStatus - The status of the assignment.

$sel:answer:Assignment', assignment_answer - The Worker's answers submitted for the HIT contained in a QuestionFormAnswers document, if the Worker provides an answer. If the Worker does not provide any answers, Answer may contain a QuestionFormAnswers document, or Answer may be empty.

$sel:submitTime:Assignment', assignment_submitTime - If the Worker has submitted results, SubmitTime is the date and time the assignment was submitted. This value is omitted from the assignment if the Worker has not yet submitted results.

$sel:workerId:Assignment', assignment_workerId - The ID of the Worker who accepted the HIT.

$sel:acceptTime:Assignment', assignment_acceptTime - The date and time the Worker accepted the assignment.

$sel:deadline:Assignment', assignment_deadline - The date and time of the deadline for the assignment. This value is derived from the deadline specification for the HIT and the date and time the Worker accepted the HIT.

$sel:approvalTime:Assignment', assignment_approvalTime - If the Worker has submitted results and the Requester has approved the results, ApprovalTime is the date and time the Requester approved the results. This value is omitted from the assignment if the Requester has not yet approved the results.

assignment_requesterFeedback :: Lens' Assignment (Maybe Text)

The feedback string included with the call to the ApproveAssignment operation or the RejectAssignment operation, if the Requester approved or rejected the assignment and specified feedback.

assignment_assignmentId :: Lens' Assignment (Maybe Text)

A unique identifier for the assignment.

assignment_hITId :: Lens' Assignment (Maybe Text)

The ID of the HIT.

assignment_autoApprovalTime :: Lens' Assignment (Maybe UTCTime)

If results have been submitted, AutoApprovalTime is the date and time the results of the assignment results are considered Approved automatically if they have not already been explicitly approved or rejected by the Requester. This value is derived from the auto-approval delay specified by the Requester in the HIT. This value is omitted from the assignment if the Worker has not yet submitted results.

assignment_rejectionTime :: Lens' Assignment (Maybe UTCTime)

If the Worker has submitted results and the Requester has rejected the results, RejectionTime is the date and time the Requester rejected the results.

assignment_assignmentStatus :: Lens' Assignment (Maybe AssignmentStatus)

The status of the assignment.

assignment_answer :: Lens' Assignment (Maybe Text)

The Worker's answers submitted for the HIT contained in a QuestionFormAnswers document, if the Worker provides an answer. If the Worker does not provide any answers, Answer may contain a QuestionFormAnswers document, or Answer may be empty.

assignment_submitTime :: Lens' Assignment (Maybe UTCTime)

If the Worker has submitted results, SubmitTime is the date and time the assignment was submitted. This value is omitted from the assignment if the Worker has not yet submitted results.

assignment_workerId :: Lens' Assignment (Maybe Text)

The ID of the Worker who accepted the HIT.

assignment_acceptTime :: Lens' Assignment (Maybe UTCTime)

The date and time the Worker accepted the assignment.

assignment_deadline :: Lens' Assignment (Maybe UTCTime)

The date and time of the deadline for the assignment. This value is derived from the deadline specification for the HIT and the date and time the Worker accepted the HIT.

assignment_approvalTime :: Lens' Assignment (Maybe UTCTime)

If the Worker has submitted results and the Requester has approved the results, ApprovalTime is the date and time the Requester approved the results. This value is omitted from the assignment if the Requester has not yet approved the results.