Network.AWS.MechanicalTurk.Types.HIT

Description

Documentation

data HIT

The HIT data structure represents a single HIT, including all the information necessary for a Worker to accept and complete the HIT.

See: newHIT smart constructor.

Constructors

HIT'

Fields

hITGroupId :: Maybe Text

The ID of the HIT Group of this HIT.

creationTime :: Maybe POSIX

The date and time the HIT was created.

assignmentDurationInSeconds :: Maybe Integer

The length of time, in seconds, that a Worker has to complete the HIT after accepting it.

autoApprovalDelayInSeconds :: Maybe Integer

The amount of time, in seconds, after the Worker submits an assignment for the HIT that the results are automatically approved by Amazon Mechanical Turk. This is the amount of time the Requester has to reject an assignment submitted by a Worker before the assignment is auto-approved and the Worker is paid.

question :: Maybe Text

The data the Worker completing the HIT uses produce the results. This is either either a QuestionForm, HTMLQuestion or an ExternalQuestion data structure.

expiration :: Maybe POSIX

The date and time the HIT expires.

hITReviewStatus :: Maybe HITReviewStatus

Indicates the review status of the HIT. Valid Values are NotReviewed | MarkedForReview | ReviewedAppropriate | ReviewedInappropriate.

hITId :: Maybe Text

A unique identifier for the HIT.

title :: Maybe Text

The title of the HIT.

hITLayoutId :: Maybe Text

The ID of the HIT Layout of this HIT.

reward :: Maybe Text

numberOfAssignmentsCompleted :: Maybe Int

The number of assignments for this HIT that have been approved or rejected.

maxAssignments :: Maybe Int

The number of times the HIT can be accepted and completed before the HIT becomes unavailable.

requesterAnnotation :: Maybe Text

An arbitrary data field the Requester who created the HIT can use. This field is visible only to the creator of the HIT.

hITStatus :: Maybe HITStatus

The status of the HIT and its assignments. Valid Values are Assignable | Unassignable | Reviewable | Reviewing | Disposed.

description :: Maybe Text

A general description of the HIT.

numberOfAssignmentsAvailable :: Maybe Int

The number of assignments for this HIT that are available for Workers to accept.

hITTypeId :: Maybe Text

The ID of the HIT type of this HIT

numberOfAssignmentsPending :: Maybe Int

The number of assignments for this HIT that are being previewed or have been accepted by Workers, but have not yet been submitted, returned, or abandoned.

qualificationRequirements :: Maybe [QualificationRequirement]

Conditions that a Worker's Qualifications must meet in order to accept the HIT. A HIT can have between zero and ten Qualification requirements. All requirements must be met in order for a Worker to accept the HIT. Additionally, other actions can be restricted using the ActionsGuarded field on each QualificationRequirement structure.

keywords :: Maybe Text

One or more words or phrases that describe the HIT, separated by commas. Search terms similar to the keywords of a HIT are more likely to have the HIT in the search results.

newHIT :: HIT

Create a value of HIT with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hITGroupId:HIT', hit_hITGroupId - The ID of the HIT Group of this HIT.

$sel:creationTime:HIT', hit_creationTime - The date and time the HIT was created.

$sel:assignmentDurationInSeconds:HIT', hit_assignmentDurationInSeconds - The length of time, in seconds, that a Worker has to complete the HIT after accepting it.

$sel:autoApprovalDelayInSeconds:HIT', hit_autoApprovalDelayInSeconds - The amount of time, in seconds, after the Worker submits an assignment for the HIT that the results are automatically approved by Amazon Mechanical Turk. This is the amount of time the Requester has to reject an assignment submitted by a Worker before the assignment is auto-approved and the Worker is paid.

$sel:question:HIT', hit_question - The data the Worker completing the HIT uses produce the results. This is either either a QuestionForm, HTMLQuestion or an ExternalQuestion data structure.

$sel:expiration:HIT', hit_expiration - The date and time the HIT expires.

$sel:hITReviewStatus:HIT', hit_hITReviewStatus - Indicates the review status of the HIT. Valid Values are NotReviewed | MarkedForReview | ReviewedAppropriate | ReviewedInappropriate.

$sel:hITId:HIT', hit_hITId - A unique identifier for the HIT.

$sel:title:HIT', hit_title - The title of the HIT.

$sel:hITLayoutId:HIT', hit_hITLayoutId - The ID of the HIT Layout of this HIT.

$sel:reward:HIT', hit_reward - Undocumented member.

$sel:numberOfAssignmentsCompleted:HIT', hit_numberOfAssignmentsCompleted - The number of assignments for this HIT that have been approved or rejected.

$sel:maxAssignments:HIT', hit_maxAssignments - The number of times the HIT can be accepted and completed before the HIT becomes unavailable.

$sel:requesterAnnotation:HIT', hit_requesterAnnotation - An arbitrary data field the Requester who created the HIT can use. This field is visible only to the creator of the HIT.

$sel:hITStatus:HIT', hit_hITStatus - The status of the HIT and its assignments. Valid Values are Assignable | Unassignable | Reviewable | Reviewing | Disposed.

$sel:description:HIT', hit_description - A general description of the HIT.

$sel:numberOfAssignmentsAvailable:HIT', hit_numberOfAssignmentsAvailable - The number of assignments for this HIT that are available for Workers to accept.

$sel:hITTypeId:HIT', hit_hITTypeId - The ID of the HIT type of this HIT

$sel:numberOfAssignmentsPending:HIT', hit_numberOfAssignmentsPending - The number of assignments for this HIT that are being previewed or have been accepted by Workers, but have not yet been submitted, returned, or abandoned.

$sel:qualificationRequirements:HIT', hit_qualificationRequirements - Conditions that a Worker's Qualifications must meet in order to accept the HIT. A HIT can have between zero and ten Qualification requirements. All requirements must be met in order for a Worker to accept the HIT. Additionally, other actions can be restricted using the ActionsGuarded field on each QualificationRequirement structure.

$sel:keywords:HIT', hit_keywords - One or more words or phrases that describe the HIT, separated by commas. Search terms similar to the keywords of a HIT are more likely to have the HIT in the search results.

hit_hITGroupId :: Lens' HIT (Maybe Text)

The ID of the HIT Group of this HIT.

hit_creationTime :: Lens' HIT (Maybe UTCTime)

The date and time the HIT was created.

hit_assignmentDurationInSeconds :: Lens' HIT (Maybe Integer)

The length of time, in seconds, that a Worker has to complete the HIT after accepting it.

hit_autoApprovalDelayInSeconds :: Lens' HIT (Maybe Integer)

The amount of time, in seconds, after the Worker submits an assignment for the HIT that the results are automatically approved by Amazon Mechanical Turk. This is the amount of time the Requester has to reject an assignment submitted by a Worker before the assignment is auto-approved and the Worker is paid.

hit_question :: Lens' HIT (Maybe Text)

The data the Worker completing the HIT uses produce the results. This is either either a QuestionForm, HTMLQuestion or an ExternalQuestion data structure.

hit_expiration :: Lens' HIT (Maybe UTCTime)

The date and time the HIT expires.

hit_hITReviewStatus :: Lens' HIT (Maybe HITReviewStatus)

Indicates the review status of the HIT. Valid Values are NotReviewed | MarkedForReview | ReviewedAppropriate | ReviewedInappropriate.

hit_hITId :: Lens' HIT (Maybe Text)

A unique identifier for the HIT.

hit_title :: Lens' HIT (Maybe Text)

The title of the HIT.

hit_hITLayoutId :: Lens' HIT (Maybe Text)

The ID of the HIT Layout of this HIT.

hit_reward :: Lens' HIT (Maybe Text)

Undocumented member.

hit_numberOfAssignmentsCompleted :: Lens' HIT (Maybe Int)

The number of assignments for this HIT that have been approved or rejected.

hit_maxAssignments :: Lens' HIT (Maybe Int)

The number of times the HIT can be accepted and completed before the HIT becomes unavailable.

hit_requesterAnnotation :: Lens' HIT (Maybe Text)

An arbitrary data field the Requester who created the HIT can use. This field is visible only to the creator of the HIT.

hit_hITStatus :: Lens' HIT (Maybe HITStatus)

The status of the HIT and its assignments. Valid Values are Assignable | Unassignable | Reviewable | Reviewing | Disposed.

hit_description :: Lens' HIT (Maybe Text)

A general description of the HIT.

hit_numberOfAssignmentsAvailable :: Lens' HIT (Maybe Int)

The number of assignments for this HIT that are available for Workers to accept.

hit_hITTypeId :: Lens' HIT (Maybe Text)

The ID of the HIT type of this HIT

hit_numberOfAssignmentsPending :: Lens' HIT (Maybe Int)

The number of assignments for this HIT that are being previewed or have been accepted by Workers, but have not yet been submitted, returned, or abandoned.

hit_qualificationRequirements :: Lens' HIT (Maybe [QualificationRequirement])

Conditions that a Worker's Qualifications must meet in order to accept the HIT. A HIT can have between zero and ten Qualification requirements. All requirements must be met in order for a Worker to accept the HIT. Additionally, other actions can be restricted using the ActionsGuarded field on each QualificationRequirement structure.

hit_keywords :: Lens' HIT (Maybe Text)

One or more words or phrases that describe the HIT, separated by commas. Search terms similar to the keywords of a HIT are more likely to have the HIT in the search results.