Network.AWS.MechanicalTurk.Types.NotificationSpecification

Description

Documentation

data NotificationSpecification

The NotificationSpecification data structure describes a HIT event notification for a HIT type.

See: newNotificationSpecification smart constructor.

Constructors

NotificationSpecification'

Fields

destination :: Text

The target for notification messages. The Destination’s format is determined by the specified Transport:

When Transport is Email, the Destination is your email address.

When Transport is SQS, the Destination is your queue URL.

When Transport is SNS, the Destination is the ARN of your topic.

transport :: NotificationTransport

The method Amazon Mechanical Turk uses to send the notification. Valid Values: Email | SQS | SNS.

version :: Text

The version of the Notification API to use. Valid value is 2006-05-05.

eventTypes :: [EventType]

The list of events that should cause notifications to be sent. Valid Values: AssignmentAccepted | AssignmentAbandoned | AssignmentReturned | AssignmentSubmitted | AssignmentRejected | AssignmentApproved | HITCreated | HITExtended | HITDisposed | HITReviewable | HITExpired | Ping. The Ping event is only valid for the SendTestEventNotification operation.

newNotificationSpecification

Arguments

:: Text
$sel:destination:NotificationSpecification'

-> NotificationTransport
$sel:transport:NotificationSpecification'

-> Text
$sel:version:NotificationSpecification'

-> NotificationSpecification

Create a value of NotificationSpecification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destination:NotificationSpecification', notificationSpecification_destination - The target for notification messages. The Destination’s format is determined by the specified Transport:

When Transport is Email, the Destination is your email address.

When Transport is SQS, the Destination is your queue URL.

When Transport is SNS, the Destination is the ARN of your topic.

$sel:transport:NotificationSpecification', notificationSpecification_transport - The method Amazon Mechanical Turk uses to send the notification. Valid Values: Email | SQS | SNS.

$sel:version:NotificationSpecification', notificationSpecification_version - The version of the Notification API to use. Valid value is 2006-05-05.

$sel:eventTypes:NotificationSpecification', notificationSpecification_eventTypes - The list of events that should cause notifications to be sent. Valid Values: AssignmentAccepted | AssignmentAbandoned | AssignmentReturned | AssignmentSubmitted | AssignmentRejected | AssignmentApproved | HITCreated | HITExtended | HITDisposed | HITReviewable | HITExpired | Ping. The Ping event is only valid for the SendTestEventNotification operation.

notificationSpecification_destination :: Lens' NotificationSpecification Text

The target for notification messages. The Destination’s format is determined by the specified Transport:

When Transport is Email, the Destination is your email address.

When Transport is SQS, the Destination is your queue URL.

When Transport is SNS, the Destination is the ARN of your topic.

notificationSpecification_transport :: Lens' NotificationSpecification NotificationTransport

The method Amazon Mechanical Turk uses to send the notification. Valid Values: Email | SQS | SNS.

notificationSpecification_version :: Lens' NotificationSpecification Text

The version of the Notification API to use. Valid value is 2006-05-05.

notificationSpecification_eventTypes :: Lens' NotificationSpecification [EventType]

The list of events that should cause notifications to be sent. Valid Values: AssignmentAccepted | AssignmentAbandoned | AssignmentReturned | AssignmentSubmitted | AssignmentRejected | AssignmentApproved | HITCreated | HITExtended | HITDisposed | HITReviewable | HITExpired | Ping. The Ping event is only valid for the SendTestEventNotification operation.