Network.AWS.MechanicalTurk.Types.Qualification

Description

Documentation

data Qualification

The Qualification data structure represents a Qualification assigned to a user, including the Qualification type and the value (score).

See: newQualification smart constructor.

Constructors

Qualification'

Fields

status :: Maybe QualificationStatus

The status of the Qualification. Valid values are Granted | Revoked.

qualificationTypeId :: Maybe Text

The ID of the Qualification type for the Qualification.

grantTime :: Maybe POSIX

The date and time the Qualification was granted to the Worker. If the Worker's Qualification was revoked, and then re-granted based on a new Qualification request, GrantTime is the date and time of the last call to the AcceptQualificationRequest operation.

localeValue :: Maybe Locale

workerId :: Maybe Text

The ID of the Worker who possesses the Qualification.

integerValue :: Maybe Int

The value (score) of the Qualification, if the Qualification has an integer value.

newQualification :: Qualification

Create a value of Qualification with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Qualification', qualification_status - The status of the Qualification. Valid values are Granted | Revoked.

$sel:qualificationTypeId:Qualification', qualification_qualificationTypeId - The ID of the Qualification type for the Qualification.

$sel:grantTime:Qualification', qualification_grantTime - The date and time the Qualification was granted to the Worker. If the Worker's Qualification was revoked, and then re-granted based on a new Qualification request, GrantTime is the date and time of the last call to the AcceptQualificationRequest operation.

$sel:localeValue:Qualification', qualification_localeValue - Undocumented member.

$sel:workerId:Qualification', qualification_workerId - The ID of the Worker who possesses the Qualification.

$sel:integerValue:Qualification', qualification_integerValue - The value (score) of the Qualification, if the Qualification has an integer value.

qualification_status :: Lens' Qualification (Maybe QualificationStatus)

The status of the Qualification. Valid values are Granted | Revoked.

qualification_qualificationTypeId :: Lens' Qualification (Maybe Text)

The ID of the Qualification type for the Qualification.

qualification_grantTime :: Lens' Qualification (Maybe UTCTime)

The date and time the Qualification was granted to the Worker. If the Worker's Qualification was revoked, and then re-granted based on a new Qualification request, GrantTime is the date and time of the last call to the AcceptQualificationRequest operation.

qualification_localeValue :: Lens' Qualification (Maybe Locale)

Undocumented member.

qualification_workerId :: Lens' Qualification (Maybe Text)

The ID of the Worker who possesses the Qualification.

qualification_integerValue :: Lens' Qualification (Maybe Int)

The value (score) of the Qualification, if the Qualification has an integer value.