Network.AWS.MechanicalTurk.Types.QualificationType

Description

Documentation

data QualificationType

The QualificationType data structure represents a Qualification type, a description of a property of a Worker that must match the requirements of a HIT for the Worker to be able to accept the HIT. The type also describes how a Worker can obtain a Qualification of that type, such as through a Qualification test.

See: newQualificationType smart constructor.

Constructors

QualificationType'

Fields

creationTime :: Maybe POSIX

The date and time the Qualification type was created.

qualificationTypeId :: Maybe Text

A unique identifier for the Qualification type. A Qualification type is given a Qualification type ID when you call the CreateQualificationType operation.

isRequestable :: Maybe Bool

Specifies whether the Qualification type is one that a user can request through the Amazon Mechanical Turk web site, such as by taking a Qualification test. This value is False for Qualifications assigned automatically by the system. Valid values are True | False.

retryDelayInSeconds :: Maybe Integer

The amount of time, in seconds, Workers must wait after taking the Qualification test before they can take it again. Workers can take a Qualification test multiple times if they were not granted the Qualification from a previous attempt, or if the test offers a gradient score and they want a better score. If not specified, retries are disabled and Workers can request a Qualification only once.

autoGranted :: Maybe Bool

Specifies that requests for the Qualification type are granted immediately, without prompting the Worker with a Qualification test. Valid values are True | False.

qualificationTypeStatus :: Maybe QualificationTypeStatus

The status of the Qualification type. A Qualification type's status determines if users can apply to receive a Qualification of this type, and if HITs can be created with requirements based on this type. Valid values are Active | Inactive.

name :: Maybe Text

The name of the Qualification type. The type name is used to identify the type, and to find the type using a Qualification type search.

testDurationInSeconds :: Maybe Integer

The amount of time, in seconds, given to a Worker to complete the Qualification test, beginning from the time the Worker requests the Qualification.

test :: Maybe Text

The questions for a Qualification test associated with this Qualification type that a user can take to obtain a Qualification of this type. This parameter must be specified if AnswerKey is present. A Qualification type cannot have both a specified Test parameter and an AutoGranted value of true.

description :: Maybe Text

A long description for the Qualification type.

autoGrantedValue :: Maybe Int

The Qualification integer value to use for automatically granted Qualifications, if AutoGranted is true. This is 1 by default.

answerKey :: Maybe Text

The answers to the Qualification test specified in the Test parameter.

keywords :: Maybe Text

One or more words or phrases that describe theQualification type, separated by commas. The Keywords make the type easier to find using a search.

newQualificationType :: QualificationType

Create a value of QualificationType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:QualificationType', qualificationType_creationTime - The date and time the Qualification type was created.

$sel:qualificationTypeId:QualificationType', qualificationType_qualificationTypeId - A unique identifier for the Qualification type. A Qualification type is given a Qualification type ID when you call the CreateQualificationType operation.

$sel:isRequestable:QualificationType', qualificationType_isRequestable - Specifies whether the Qualification type is one that a user can request through the Amazon Mechanical Turk web site, such as by taking a Qualification test. This value is False for Qualifications assigned automatically by the system. Valid values are True | False.

$sel:retryDelayInSeconds:QualificationType', qualificationType_retryDelayInSeconds - The amount of time, in seconds, Workers must wait after taking the Qualification test before they can take it again. Workers can take a Qualification test multiple times if they were not granted the Qualification from a previous attempt, or if the test offers a gradient score and they want a better score. If not specified, retries are disabled and Workers can request a Qualification only once.

$sel:autoGranted:QualificationType', qualificationType_autoGranted - Specifies that requests for the Qualification type are granted immediately, without prompting the Worker with a Qualification test. Valid values are True | False.

$sel:qualificationTypeStatus:QualificationType', qualificationType_qualificationTypeStatus - The status of the Qualification type. A Qualification type's status determines if users can apply to receive a Qualification of this type, and if HITs can be created with requirements based on this type. Valid values are Active | Inactive.

$sel:name:QualificationType', qualificationType_name - The name of the Qualification type. The type name is used to identify the type, and to find the type using a Qualification type search.

$sel:testDurationInSeconds:QualificationType', qualificationType_testDurationInSeconds - The amount of time, in seconds, given to a Worker to complete the Qualification test, beginning from the time the Worker requests the Qualification.

$sel:test:QualificationType', qualificationType_test - The questions for a Qualification test associated with this Qualification type that a user can take to obtain a Qualification of this type. This parameter must be specified if AnswerKey is present. A Qualification type cannot have both a specified Test parameter and an AutoGranted value of true.

$sel:description:QualificationType', qualificationType_description - A long description for the Qualification type.

$sel:autoGrantedValue:QualificationType', qualificationType_autoGrantedValue - The Qualification integer value to use for automatically granted Qualifications, if AutoGranted is true. This is 1 by default.

$sel:answerKey:QualificationType', qualificationType_answerKey - The answers to the Qualification test specified in the Test parameter.

$sel:keywords:QualificationType', qualificationType_keywords - One or more words or phrases that describe theQualification type, separated by commas. The Keywords make the type easier to find using a search.

qualificationType_creationTime :: Lens' QualificationType (Maybe UTCTime)

The date and time the Qualification type was created.

qualificationType_qualificationTypeId :: Lens' QualificationType (Maybe Text)

A unique identifier for the Qualification type. A Qualification type is given a Qualification type ID when you call the CreateQualificationType operation.

qualificationType_isRequestable :: Lens' QualificationType (Maybe Bool)

Specifies whether the Qualification type is one that a user can request through the Amazon Mechanical Turk web site, such as by taking a Qualification test. This value is False for Qualifications assigned automatically by the system. Valid values are True | False.

qualificationType_retryDelayInSeconds :: Lens' QualificationType (Maybe Integer)

The amount of time, in seconds, Workers must wait after taking the Qualification test before they can take it again. Workers can take a Qualification test multiple times if they were not granted the Qualification from a previous attempt, or if the test offers a gradient score and they want a better score. If not specified, retries are disabled and Workers can request a Qualification only once.

qualificationType_autoGranted :: Lens' QualificationType (Maybe Bool)

Specifies that requests for the Qualification type are granted immediately, without prompting the Worker with a Qualification test. Valid values are True | False.

qualificationType_qualificationTypeStatus :: Lens' QualificationType (Maybe QualificationTypeStatus)

The status of the Qualification type. A Qualification type's status determines if users can apply to receive a Qualification of this type, and if HITs can be created with requirements based on this type. Valid values are Active | Inactive.

qualificationType_name :: Lens' QualificationType (Maybe Text)

The name of the Qualification type. The type name is used to identify the type, and to find the type using a Qualification type search.

qualificationType_testDurationInSeconds :: Lens' QualificationType (Maybe Integer)

The amount of time, in seconds, given to a Worker to complete the Qualification test, beginning from the time the Worker requests the Qualification.

qualificationType_test :: Lens' QualificationType (Maybe Text)

The questions for a Qualification test associated with this Qualification type that a user can take to obtain a Qualification of this type. This parameter must be specified if AnswerKey is present. A Qualification type cannot have both a specified Test parameter and an AutoGranted value of true.

qualificationType_description :: Lens' QualificationType (Maybe Text)

A long description for the Qualification type.

qualificationType_autoGrantedValue :: Lens' QualificationType (Maybe Int)

The Qualification integer value to use for automatically granted Qualifications, if AutoGranted is true. This is 1 by default.

qualificationType_answerKey :: Lens' QualificationType (Maybe Text)

The answers to the Qualification test specified in the Test parameter.

qualificationType_keywords :: Lens' QualificationType (Maybe Text)

One or more words or phrases that describe theQualification type, separated by commas. The Keywords make the type easier to find using a search.