Network.AWS.MechanicalTurk.Types.ReviewResultDetail

Description

Documentation

data ReviewResultDetail

This data structure is returned multiple times for each result specified in the Review Policy.

See: newReviewResultDetail smart constructor.

Constructors

ReviewResultDetail'

Fields

key :: Maybe Text

Key identifies the particular piece of reviewed information.

subjectType :: Maybe Text

The type of the object from the SubjectId field.

subjectId :: Maybe Text

The HITID or AssignmentId about which this result was taken. Note that HIT-level Review Policies will often emit results about both the HIT itself and its Assignments, while Assignment-level review policies generally only emit results about the Assignment itself.

actionId :: Maybe Text

A unique identifier of the Review action result.

value :: Maybe Text

The values of Key provided by the review policies you have selected.

questionId :: Maybe Text

Specifies the QuestionId the result is describing. Depending on whether the TargetType is a HIT or Assignment this results could specify multiple values. If TargetType is HIT and QuestionId is absent, then the result describes results of the HIT, including the HIT agreement score. If ObjectType is Assignment and QuestionId is absent, then the result describes the Worker's performance on the HIT.

newReviewResultDetail :: ReviewResultDetail

Create a value of ReviewResultDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:ReviewResultDetail', reviewResultDetail_key - Key identifies the particular piece of reviewed information.

$sel:subjectType:ReviewResultDetail', reviewResultDetail_subjectType - The type of the object from the SubjectId field.

$sel:subjectId:ReviewResultDetail', reviewResultDetail_subjectId - The HITID or AssignmentId about which this result was taken. Note that HIT-level Review Policies will often emit results about both the HIT itself and its Assignments, while Assignment-level review policies generally only emit results about the Assignment itself.

$sel:actionId:ReviewResultDetail', reviewResultDetail_actionId - A unique identifier of the Review action result.

$sel:value:ReviewResultDetail', reviewResultDetail_value - The values of Key provided by the review policies you have selected.

$sel:questionId:ReviewResultDetail', reviewResultDetail_questionId - Specifies the QuestionId the result is describing. Depending on whether the TargetType is a HIT or Assignment this results could specify multiple values. If TargetType is HIT and QuestionId is absent, then the result describes results of the HIT, including the HIT agreement score. If ObjectType is Assignment and QuestionId is absent, then the result describes the Worker's performance on the HIT.

reviewResultDetail_key :: Lens' ReviewResultDetail (Maybe Text)

Key identifies the particular piece of reviewed information.

reviewResultDetail_subjectType :: Lens' ReviewResultDetail (Maybe Text)

The type of the object from the SubjectId field.

reviewResultDetail_subjectId :: Lens' ReviewResultDetail (Maybe Text)

The HITID or AssignmentId about which this result was taken. Note that HIT-level Review Policies will often emit results about both the HIT itself and its Assignments, while Assignment-level review policies generally only emit results about the Assignment itself.

reviewResultDetail_actionId :: Lens' ReviewResultDetail (Maybe Text)

A unique identifier of the Review action result.

reviewResultDetail_value :: Lens' ReviewResultDetail (Maybe Text)

The values of Key provided by the review policies you have selected.

reviewResultDetail_questionId :: Lens' ReviewResultDetail (Maybe Text)

Specifies the QuestionId the result is describing. Depending on whether the TargetType is a HIT or Assignment this results could specify multiple values. If TargetType is HIT and QuestionId is absent, then the result describes results of the HIT, including the HIT agreement score. If ObjectType is Assignment and QuestionId is absent, then the result describes the Worker's performance on the HIT.