Network.AWS.MechanicalTurk.UpdateNotificationSettings

Description

The UpdateNotificationSettings operation creates, updates, disables or re-enables notifications for a HIT type. If you call the UpdateNotificationSettings operation for a HIT type that already has a notification specification, the operation replaces the old specification with a new one. You can call the UpdateNotificationSettings operation to enable or disable notifications for the HIT type, without having to modify the notification specification itself by providing updates to the Active status without specifying a new notification specification. To change the Active status of a HIT type's notifications, the HIT type must already have a notification specification, or one must be provided in the same call to UpdateNotificationSettings.

Creating a Request

data UpdateNotificationSettings

See: newUpdateNotificationSettings smart constructor.

Constructors

UpdateNotificationSettings'

Fields

active :: Maybe Bool

Specifies whether notifications are sent for HITs of this HIT type, according to the notification specification. You must specify either the Notification parameter or the Active parameter for the call to UpdateNotificationSettings to succeed.

notification :: Maybe NotificationSpecification

The notification specification for the HIT type.

hITTypeId :: Text

The ID of the HIT type whose notification specification is being updated.

newUpdateNotificationSettings

Arguments

:: Text
$sel:hITTypeId:UpdateNotificationSettings'

-> UpdateNotificationSettings

Create a value of UpdateNotificationSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:active:UpdateNotificationSettings', updateNotificationSettings_active - Specifies whether notifications are sent for HITs of this HIT type, according to the notification specification. You must specify either the Notification parameter or the Active parameter for the call to UpdateNotificationSettings to succeed.

$sel:notification:UpdateNotificationSettings', updateNotificationSettings_notification - The notification specification for the HIT type.

$sel:hITTypeId:UpdateNotificationSettings', updateNotificationSettings_hITTypeId - The ID of the HIT type whose notification specification is being updated.

Request Lenses

updateNotificationSettings_active :: Lens' UpdateNotificationSettings (Maybe Bool)

Specifies whether notifications are sent for HITs of this HIT type, according to the notification specification. You must specify either the Notification parameter or the Active parameter for the call to UpdateNotificationSettings to succeed.

updateNotificationSettings_notification :: Lens' UpdateNotificationSettings (Maybe NotificationSpecification)

The notification specification for the HIT type.

updateNotificationSettings_hITTypeId :: Lens' UpdateNotificationSettings Text

The ID of the HIT type whose notification specification is being updated.

Destructuring the Response

data UpdateNotificationSettingsResponse

See: newUpdateNotificationSettingsResponse smart constructor.

Constructors

UpdateNotificationSettingsResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateNotificationSettingsResponse

Arguments

:: Int
$sel:httpStatus:UpdateNotificationSettingsResponse'

-> UpdateNotificationSettingsResponse

Create a value of UpdateNotificationSettingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateNotificationSettingsResponse', updateNotificationSettingsResponse_httpStatus - The response's http status code.

Response Lenses

updateNotificationSettingsResponse_httpStatus :: Lens' UpdateNotificationSettingsResponse Int

The response's http status code.