Network.AWS.MechanicalTurk.UpdateQualificationType

Description

The UpdateQualificationType operation modifies the attributes of an existing Qualification type, which is represented by a QualificationType data structure. Only the owner of a Qualification type can modify its attributes.

Most attributes of a Qualification type can be changed after the type has been created. However, the Name and Keywords fields cannot be modified. The RetryDelayInSeconds parameter can be modified or added to change the delay or to enable retries, but RetryDelayInSeconds cannot be used to disable retries.

You can use this operation to update the test for a Qualification type. The test is updated based on the values specified for the Test, TestDurationInSeconds and AnswerKey parameters. All three parameters specify the updated test. If you are updating the test for a type, you must specify the Test and TestDurationInSeconds parameters. The AnswerKey parameter is optional; omitting it specifies that the updated test does not have an answer key.

If you omit the Test parameter, the test for the Qualification type is unchanged. There is no way to remove a test from a Qualification type that has one. If the type already has a test, you cannot update it to be AutoGranted. If the Qualification type does not have a test and one is provided by an update, the type will henceforth have a test.

If you want to update the test duration or answer key for an existing test without changing the questions, you must specify a Test parameter with the original questions, along with the updated values.

If you provide an updated Test but no AnswerKey, the new test will not have an answer key. Requests for such Qualifications must be granted manually.

You can also update the AutoGranted and AutoGrantedValue attributes of the Qualification type.

Creating a Request

data UpdateQualificationType

See: newUpdateQualificationType smart constructor.

Constructors

UpdateQualificationType'

Fields

retryDelayInSeconds :: Maybe Integer

The amount of time, in seconds, that Workers must wait after requesting a Qualification of the specified Qualification type before they can retry the Qualification request. It is not possible to disable retries for a Qualification type after it has been created with retries enabled. If you want to disable retries, you must dispose of the existing retry-enabled Qualification type using DisposeQualificationType and then create a new Qualification type with retries disabled using CreateQualificationType.

autoGranted :: Maybe Bool

Specifies whether requests for the Qualification type are granted immediately, without prompting the Worker with a Qualification test.

Constraints: If the Test parameter is specified, this parameter cannot be true.

qualificationTypeStatus :: Maybe QualificationTypeStatus

The new status of the Qualification type - Active | Inactive

testDurationInSeconds :: Maybe Integer

The number of seconds the Worker has to complete the Qualification test, starting from the time the Worker requests the Qualification.

test :: Maybe Text

The questions for the Qualification test a Worker must answer correctly to obtain a Qualification of this type. If this parameter is specified, TestDurationInSeconds must also be specified.

Constraints: Must not be longer than 65535 bytes. Must be a QuestionForm data structure. This parameter cannot be specified if AutoGranted is true.

Constraints: None. If not specified, the Worker may request the Qualification without answering any questions.

description :: Maybe Text

The new description of the Qualification type.

autoGrantedValue :: Maybe Int

The Qualification value to use for automatically granted Qualifications. This parameter is used only if the AutoGranted parameter is true.

answerKey :: Maybe Text

The answers to the Qualification test specified in the Test parameter, in the form of an AnswerKey data structure.

qualificationTypeId :: Text

The ID of the Qualification type to update.

newUpdateQualificationType

Arguments

:: Text
$sel:qualificationTypeId:UpdateQualificationType'

-> UpdateQualificationType

Create a value of UpdateQualificationType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retryDelayInSeconds:UpdateQualificationType', updateQualificationType_retryDelayInSeconds - The amount of time, in seconds, that Workers must wait after requesting a Qualification of the specified Qualification type before they can retry the Qualification request. It is not possible to disable retries for a Qualification type after it has been created with retries enabled. If you want to disable retries, you must dispose of the existing retry-enabled Qualification type using DisposeQualificationType and then create a new Qualification type with retries disabled using CreateQualificationType.

$sel:autoGranted:UpdateQualificationType', updateQualificationType_autoGranted - Specifies whether requests for the Qualification type are granted immediately, without prompting the Worker with a Qualification test.

Constraints: If the Test parameter is specified, this parameter cannot be true.

$sel:qualificationTypeStatus:UpdateQualificationType', updateQualificationType_qualificationTypeStatus - The new status of the Qualification type - Active | Inactive

$sel:testDurationInSeconds:UpdateQualificationType', updateQualificationType_testDurationInSeconds - The number of seconds the Worker has to complete the Qualification test, starting from the time the Worker requests the Qualification.

$sel:test:UpdateQualificationType', updateQualificationType_test - The questions for the Qualification test a Worker must answer correctly to obtain a Qualification of this type. If this parameter is specified, TestDurationInSeconds must also be specified.

Constraints: Must not be longer than 65535 bytes. Must be a QuestionForm data structure. This parameter cannot be specified if AutoGranted is true.

Constraints: None. If not specified, the Worker may request the Qualification without answering any questions.

$sel:description:UpdateQualificationType', updateQualificationType_description - The new description of the Qualification type.

$sel:autoGrantedValue:UpdateQualificationType', updateQualificationType_autoGrantedValue - The Qualification value to use for automatically granted Qualifications. This parameter is used only if the AutoGranted parameter is true.

$sel:answerKey:UpdateQualificationType', updateQualificationType_answerKey - The answers to the Qualification test specified in the Test parameter, in the form of an AnswerKey data structure.

$sel:qualificationTypeId:UpdateQualificationType', updateQualificationType_qualificationTypeId - The ID of the Qualification type to update.

Request Lenses

updateQualificationType_retryDelayInSeconds :: Lens' UpdateQualificationType (Maybe Integer)

The amount of time, in seconds, that Workers must wait after requesting a Qualification of the specified Qualification type before they can retry the Qualification request. It is not possible to disable retries for a Qualification type after it has been created with retries enabled. If you want to disable retries, you must dispose of the existing retry-enabled Qualification type using DisposeQualificationType and then create a new Qualification type with retries disabled using CreateQualificationType.

updateQualificationType_autoGranted :: Lens' UpdateQualificationType (Maybe Bool)

Specifies whether requests for the Qualification type are granted immediately, without prompting the Worker with a Qualification test.

Constraints: If the Test parameter is specified, this parameter cannot be true.

updateQualificationType_qualificationTypeStatus :: Lens' UpdateQualificationType (Maybe QualificationTypeStatus)

The new status of the Qualification type - Active | Inactive

updateQualificationType_testDurationInSeconds :: Lens' UpdateQualificationType (Maybe Integer)

The number of seconds the Worker has to complete the Qualification test, starting from the time the Worker requests the Qualification.

updateQualificationType_test :: Lens' UpdateQualificationType (Maybe Text)

The questions for the Qualification test a Worker must answer correctly to obtain a Qualification of this type. If this parameter is specified, TestDurationInSeconds must also be specified.

Constraints: Must not be longer than 65535 bytes. Must be a QuestionForm data structure. This parameter cannot be specified if AutoGranted is true.

Constraints: None. If not specified, the Worker may request the Qualification without answering any questions.

updateQualificationType_description :: Lens' UpdateQualificationType (Maybe Text)

The new description of the Qualification type.

updateQualificationType_autoGrantedValue :: Lens' UpdateQualificationType (Maybe Int)

The Qualification value to use for automatically granted Qualifications. This parameter is used only if the AutoGranted parameter is true.

updateQualificationType_answerKey :: Lens' UpdateQualificationType (Maybe Text)

The answers to the Qualification test specified in the Test parameter, in the form of an AnswerKey data structure.

updateQualificationType_qualificationTypeId :: Lens' UpdateQualificationType Text

The ID of the Qualification type to update.

Destructuring the Response

data UpdateQualificationTypeResponse

See: newUpdateQualificationTypeResponse smart constructor.

Constructors

UpdateQualificationTypeResponse'

Fields

qualificationType :: Maybe QualificationType

Contains a QualificationType data structure.

httpStatus :: Int

The response's http status code.

newUpdateQualificationTypeResponse

Arguments

:: Int
$sel:httpStatus:UpdateQualificationTypeResponse'

-> UpdateQualificationTypeResponse

Create a value of UpdateQualificationTypeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:qualificationType:UpdateQualificationTypeResponse', updateQualificationTypeResponse_qualificationType - Contains a QualificationType data structure.

$sel:httpStatus:UpdateQualificationTypeResponse', updateQualificationTypeResponse_httpStatus - The response's http status code.

Response Lenses

updateQualificationTypeResponse_qualificationType :: Lens' UpdateQualificationTypeResponse (Maybe QualificationType)

Contains a QualificationType data structure.

updateQualificationTypeResponse_httpStatus :: Lens' UpdateQualificationTypeResponse Int

The response's http status code.