Network.AWS.MediaConnect.AddFlowMediaStreams

Description

Adds media streams to an existing flow. After you add a media stream to a flow, you can associate it with a source and/or an output that uses the ST 2110 JPEG XS or CDI protocol.

Creating a Request

data AddFlowMediaStreams

A request to add media streams to the flow.

See: newAddFlowMediaStreams smart constructor.

Constructors

AddFlowMediaStreams'

Fields

flowArn :: Text

The Amazon Resource Name (ARN) of the flow.

mediaStreams :: [AddMediaStreamRequest]

The media streams that you want to add to the flow.

newAddFlowMediaStreams

Arguments

:: Text
$sel:flowArn:AddFlowMediaStreams'

-> AddFlowMediaStreams

Create a value of AddFlowMediaStreams with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flowArn:AddFlowMediaStreams', addFlowMediaStreams_flowArn - The Amazon Resource Name (ARN) of the flow.

$sel:mediaStreams:AddFlowMediaStreams', addFlowMediaStreams_mediaStreams - The media streams that you want to add to the flow.

Request Lenses

addFlowMediaStreams_flowArn :: Lens' AddFlowMediaStreams Text

The Amazon Resource Name (ARN) of the flow.

addFlowMediaStreams_mediaStreams :: Lens' AddFlowMediaStreams [AddMediaStreamRequest]

The media streams that you want to add to the flow.

Destructuring the Response

data AddFlowMediaStreamsResponse

See: newAddFlowMediaStreamsResponse smart constructor.

Constructors

AddFlowMediaStreamsResponse'

Fields

flowArn :: Maybe Text

The ARN of the flow that you added media streams to.

mediaStreams :: Maybe [MediaStream]

The media streams that you added to the flow.

httpStatus :: Int

The response's http status code.

newAddFlowMediaStreamsResponse

Arguments

:: Int
$sel:httpStatus:AddFlowMediaStreamsResponse'

-> AddFlowMediaStreamsResponse

Create a value of AddFlowMediaStreamsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flowArn:AddFlowMediaStreams', addFlowMediaStreamsResponse_flowArn - The ARN of the flow that you added media streams to.

$sel:mediaStreams:AddFlowMediaStreams', addFlowMediaStreamsResponse_mediaStreams - The media streams that you added to the flow.

$sel:httpStatus:AddFlowMediaStreamsResponse', addFlowMediaStreamsResponse_httpStatus - The response's http status code.

Response Lenses

addFlowMediaStreamsResponse_flowArn :: Lens' AddFlowMediaStreamsResponse (Maybe Text)

The ARN of the flow that you added media streams to.

addFlowMediaStreamsResponse_mediaStreams :: Lens' AddFlowMediaStreamsResponse (Maybe [MediaStream])

The media streams that you added to the flow.

addFlowMediaStreamsResponse_httpStatus :: Lens' AddFlowMediaStreamsResponse Int

The response's http status code.