Network.AWS.MediaConnect.AddFlowVpcInterfaces

Description

Adds VPC interfaces to flow

Creating a Request

data AddFlowVpcInterfaces

A request to add VPC interfaces to the flow.

See: newAddFlowVpcInterfaces smart constructor.

Constructors

AddFlowVpcInterfaces'

Fields

flowArn :: Text

The flow that you want to mutate.

vpcInterfaces :: [VpcInterfaceRequest]

A list of VPC interfaces that you want to add.

newAddFlowVpcInterfaces

Arguments

:: Text
$sel:flowArn:AddFlowVpcInterfaces'

-> AddFlowVpcInterfaces

Create a value of AddFlowVpcInterfaces with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flowArn:AddFlowVpcInterfaces', addFlowVpcInterfaces_flowArn - The flow that you want to mutate.

$sel:vpcInterfaces:AddFlowVpcInterfaces', addFlowVpcInterfaces_vpcInterfaces - A list of VPC interfaces that you want to add.

Request Lenses

addFlowVpcInterfaces_flowArn :: Lens' AddFlowVpcInterfaces Text

The flow that you want to mutate.

addFlowVpcInterfaces_vpcInterfaces :: Lens' AddFlowVpcInterfaces [VpcInterfaceRequest]

A list of VPC interfaces that you want to add.

Destructuring the Response

data AddFlowVpcInterfacesResponse

See: newAddFlowVpcInterfacesResponse smart constructor.

Constructors

AddFlowVpcInterfacesResponse'

Fields

flowArn :: Maybe Text

The ARN of the flow that these VPC interfaces were added to.

vpcInterfaces :: Maybe [VpcInterface]

The details of the newly added VPC interfaces.

httpStatus :: Int

The response's http status code.

newAddFlowVpcInterfacesResponse

Arguments

:: Int
$sel:httpStatus:AddFlowVpcInterfacesResponse'

-> AddFlowVpcInterfacesResponse

Create a value of AddFlowVpcInterfacesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flowArn:AddFlowVpcInterfaces', addFlowVpcInterfacesResponse_flowArn - The ARN of the flow that these VPC interfaces were added to.

$sel:vpcInterfaces:AddFlowVpcInterfaces', addFlowVpcInterfacesResponse_vpcInterfaces - The details of the newly added VPC interfaces.

$sel:httpStatus:AddFlowVpcInterfacesResponse', addFlowVpcInterfacesResponse_httpStatus - The response's http status code.

Response Lenses

addFlowVpcInterfacesResponse_flowArn :: Lens' AddFlowVpcInterfacesResponse (Maybe Text)

The ARN of the flow that these VPC interfaces were added to.

addFlowVpcInterfacesResponse_vpcInterfaces :: Lens' AddFlowVpcInterfacesResponse (Maybe [VpcInterface])

The details of the newly added VPC interfaces.

addFlowVpcInterfacesResponse_httpStatus :: Lens' AddFlowVpcInterfacesResponse Int

The response's http status code.