Network.AWS.MediaConnect.CreateFlow

Description

Creates a new flow. The request must include one source. The request optionally can include outputs (up to 50) and entitlements (up to 50).

Creating a Request

data CreateFlow

Creates a new flow. The request must include one source. The request optionally can include outputs (up to 50) and entitlements (up to 50).

See: newCreateFlow smart constructor.

Constructors

CreateFlow'

Fields

mediaStreams :: Maybe [AddMediaStreamRequest]

The media streams that you want to add to the flow. You can associate these media streams with sources and outputs on the flow.

sourceFailoverConfig :: Maybe FailoverConfig

vpcInterfaces :: Maybe [VpcInterfaceRequest]

The VPC interfaces you want on the flow.

sources :: Maybe [SetSourceRequest]

outputs :: Maybe [AddOutputRequest]

The outputs that you want to add to this flow.

availabilityZone :: Maybe Text

The Availability Zone that you want to create the flow in. These options are limited to the Availability Zones within the current AWS Region.

entitlements :: Maybe [GrantEntitlementRequest]

The entitlements that you want to grant on a flow.

source :: Maybe SetSourceRequest

name :: Text

The name of the flow.

newCreateFlow

Arguments

:: Text
$sel:name:CreateFlow'

-> CreateFlow

Create a value of CreateFlow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mediaStreams:CreateFlow', createFlow_mediaStreams - The media streams that you want to add to the flow. You can associate these media streams with sources and outputs on the flow.

$sel:sourceFailoverConfig:CreateFlow', createFlow_sourceFailoverConfig - Undocumented member.

$sel:vpcInterfaces:CreateFlow', createFlow_vpcInterfaces - The VPC interfaces you want on the flow.

$sel:sources:CreateFlow', createFlow_sources - Undocumented member.

$sel:outputs:CreateFlow', createFlow_outputs - The outputs that you want to add to this flow.

$sel:availabilityZone:CreateFlow', createFlow_availabilityZone - The Availability Zone that you want to create the flow in. These options are limited to the Availability Zones within the current AWS Region.

$sel:entitlements:CreateFlow', createFlow_entitlements - The entitlements that you want to grant on a flow.

$sel:source:CreateFlow', createFlow_source - Undocumented member.

$sel:name:CreateFlow', createFlow_name - The name of the flow.

Request Lenses

createFlow_mediaStreams :: Lens' CreateFlow (Maybe [AddMediaStreamRequest])

The media streams that you want to add to the flow. You can associate these media streams with sources and outputs on the flow.

createFlow_sourceFailoverConfig :: Lens' CreateFlow (Maybe FailoverConfig)

Undocumented member.

createFlow_vpcInterfaces :: Lens' CreateFlow (Maybe [VpcInterfaceRequest])

The VPC interfaces you want on the flow.

createFlow_sources :: Lens' CreateFlow (Maybe [SetSourceRequest])

Undocumented member.

createFlow_outputs :: Lens' CreateFlow (Maybe [AddOutputRequest])

The outputs that you want to add to this flow.

createFlow_availabilityZone :: Lens' CreateFlow (Maybe Text)

The Availability Zone that you want to create the flow in. These options are limited to the Availability Zones within the current AWS Region.

createFlow_entitlements :: Lens' CreateFlow (Maybe [GrantEntitlementRequest])

The entitlements that you want to grant on a flow.

createFlow_source :: Lens' CreateFlow (Maybe SetSourceRequest)

Undocumented member.

createFlow_name :: Lens' CreateFlow Text

The name of the flow.

Destructuring the Response

data CreateFlowResponse

See: newCreateFlowResponse smart constructor.

Constructors

CreateFlowResponse'

Fields

flow :: Maybe Flow

httpStatus :: Int

The response's http status code.

newCreateFlowResponse

Arguments

:: Int
$sel:httpStatus:CreateFlowResponse'

-> CreateFlowResponse

Create a value of CreateFlowResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flow:CreateFlowResponse', createFlowResponse_flow - Undocumented member.

$sel:httpStatus:CreateFlowResponse', createFlowResponse_httpStatus - The response's http status code.

Response Lenses

createFlowResponse_flow :: Lens' CreateFlowResponse (Maybe Flow)

Undocumented member.

createFlowResponse_httpStatus :: Lens' CreateFlowResponse Int

The response's http status code.