Network.AWS.MediaConnect.ListFlows

Description

Displays a list of flows that are associated with this account. This request returns a paginated result.

This operation returns paginated results.

Creating a Request

data ListFlows

See: newListFlows smart constructor.

Constructors

ListFlows'

Fields

nextToken :: Maybe Text

The token that identifies which batch of results that you want to see. For example, you submit a ListFlows request with MaxResults set at 5. The service returns the first batch of results (up to 5) and a NextToken value. To see the next batch of results, you can submit the ListFlows request a second time and specify the NextToken value.

maxResults :: Maybe Natural

The maximum number of results to return per API request. For example, you submit a ListFlows request with MaxResults set at 5. Although 20 items match your request, the service returns no more than the first 5 items. (The service also returns a NextToken value that you can use to fetch the next batch of results.) The service might return fewer results than the MaxResults value. If MaxResults is not included in the request, the service defaults to pagination with a maximum of 10 results per page.

newListFlows :: ListFlows

Create a value of ListFlows with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFlows', listFlows_nextToken - The token that identifies which batch of results that you want to see. For example, you submit a ListFlows request with MaxResults set at 5. The service returns the first batch of results (up to 5) and a NextToken value. To see the next batch of results, you can submit the ListFlows request a second time and specify the NextToken value.

$sel:maxResults:ListFlows', listFlows_maxResults - The maximum number of results to return per API request. For example, you submit a ListFlows request with MaxResults set at 5. Although 20 items match your request, the service returns no more than the first 5 items. (The service also returns a NextToken value that you can use to fetch the next batch of results.) The service might return fewer results than the MaxResults value. If MaxResults is not included in the request, the service defaults to pagination with a maximum of 10 results per page.

Request Lenses

listFlows_nextToken :: Lens' ListFlows (Maybe Text)

The token that identifies which batch of results that you want to see. For example, you submit a ListFlows request with MaxResults set at 5. The service returns the first batch of results (up to 5) and a NextToken value. To see the next batch of results, you can submit the ListFlows request a second time and specify the NextToken value.

listFlows_maxResults :: Lens' ListFlows (Maybe Natural)

The maximum number of results to return per API request. For example, you submit a ListFlows request with MaxResults set at 5. Although 20 items match your request, the service returns no more than the first 5 items. (The service also returns a NextToken value that you can use to fetch the next batch of results.) The service might return fewer results than the MaxResults value. If MaxResults is not included in the request, the service defaults to pagination with a maximum of 10 results per page.

Destructuring the Response

data ListFlowsResponse

See: newListFlowsResponse smart constructor.

Constructors

ListFlowsResponse'

Fields

nextToken :: Maybe Text

The token that identifies which batch of results that you want to see. For example, you submit a ListFlows request with MaxResults set at 5. The service returns the first batch of results (up to 5) and a NextToken value. To see the next batch of results, you can submit the ListFlows request a second time and specify the NextToken value.

flows :: Maybe [ListedFlow]

A list of flow summaries.

httpStatus :: Int

The response's http status code.

newListFlowsResponse

Arguments

:: Int
$sel:httpStatus:ListFlowsResponse'

-> ListFlowsResponse

Create a value of ListFlowsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFlows', listFlowsResponse_nextToken - The token that identifies which batch of results that you want to see. For example, you submit a ListFlows request with MaxResults set at 5. The service returns the first batch of results (up to 5) and a NextToken value. To see the next batch of results, you can submit the ListFlows request a second time and specify the NextToken value.

$sel:flows:ListFlowsResponse', listFlowsResponse_flows - A list of flow summaries.

$sel:httpStatus:ListFlowsResponse', listFlowsResponse_httpStatus - The response's http status code.

Response Lenses

listFlowsResponse_nextToken :: Lens' ListFlowsResponse (Maybe Text)

The token that identifies which batch of results that you want to see. For example, you submit a ListFlows request with MaxResults set at 5. The service returns the first batch of results (up to 5) and a NextToken value. To see the next batch of results, you can submit the ListFlows request a second time and specify the NextToken value.

listFlowsResponse_flows :: Lens' ListFlowsResponse (Maybe [ListedFlow])

A list of flow summaries.

listFlowsResponse_httpStatus :: Lens' ListFlowsResponse Int

The response's http status code.