Network.AWS.MediaConnect.PurchaseOffering

Description

Submits a request to purchase an offering. If you already have an active reservation, you can't purchase another offering.

Creating a Request

data PurchaseOffering

A request to purchase a offering.

See: newPurchaseOffering smart constructor.

Constructors

PurchaseOffering'

Fields

offeringArn :: Text

The Amazon Resource Name (ARN) of the offering.

start :: Text

The date and time that you want the reservation to begin, in Coordinated Universal Time (UTC). You can specify any date and time between 12:00am on the first day of the current month to the current time on today's date, inclusive. Specify the start in a 24-hour notation. Use the following format: YYYY-MM-DDTHH:mm:SSZ, where T and Z are literal characters. For example, to specify 11:30pm on March 5, 2020, enter 2020-03-05T23:30:00Z.

reservationName :: Text

The name that you want to use for the reservation.

newPurchaseOffering

Arguments

:: Text
$sel:offeringArn:PurchaseOffering'

-> Text
$sel:start:PurchaseOffering'

-> Text
$sel:reservationName:PurchaseOffering'

-> PurchaseOffering

Create a value of PurchaseOffering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:offeringArn:PurchaseOffering', purchaseOffering_offeringArn - The Amazon Resource Name (ARN) of the offering.

$sel:start:PurchaseOffering', purchaseOffering_start - The date and time that you want the reservation to begin, in Coordinated Universal Time (UTC). You can specify any date and time between 12:00am on the first day of the current month to the current time on today's date, inclusive. Specify the start in a 24-hour notation. Use the following format: YYYY-MM-DDTHH:mm:SSZ, where T and Z are literal characters. For example, to specify 11:30pm on March 5, 2020, enter 2020-03-05T23:30:00Z.

$sel:reservationName:PurchaseOffering', purchaseOffering_reservationName - The name that you want to use for the reservation.

Request Lenses

purchaseOffering_offeringArn :: Lens' PurchaseOffering Text

The Amazon Resource Name (ARN) of the offering.

purchaseOffering_start :: Lens' PurchaseOffering Text

The date and time that you want the reservation to begin, in Coordinated Universal Time (UTC). You can specify any date and time between 12:00am on the first day of the current month to the current time on today's date, inclusive. Specify the start in a 24-hour notation. Use the following format: YYYY-MM-DDTHH:mm:SSZ, where T and Z are literal characters. For example, to specify 11:30pm on March 5, 2020, enter 2020-03-05T23:30:00Z.

purchaseOffering_reservationName :: Lens' PurchaseOffering Text

The name that you want to use for the reservation.

Destructuring the Response

data PurchaseOfferingResponse

See: newPurchaseOfferingResponse smart constructor.

Constructors

PurchaseOfferingResponse'

Fields

reservation :: Maybe Reservation

httpStatus :: Int

The response's http status code.

newPurchaseOfferingResponse

Arguments

:: Int
$sel:httpStatus:PurchaseOfferingResponse'

-> PurchaseOfferingResponse

Create a value of PurchaseOfferingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reservation:PurchaseOfferingResponse', purchaseOfferingResponse_reservation - Undocumented member.

$sel:httpStatus:PurchaseOfferingResponse', purchaseOfferingResponse_httpStatus - The response's http status code.

Response Lenses

purchaseOfferingResponse_reservation :: Lens' PurchaseOfferingResponse (Maybe Reservation)

Undocumented member.

purchaseOfferingResponse_httpStatus :: Lens' PurchaseOfferingResponse Int

The response's http status code.