Network.AWS.MediaConnect.RemoveFlowVpcInterface

Description

Removes a VPC Interface from an existing flow. This request can be made only on a VPC interface that does not have a Source or Output associated with it. If the VPC interface is referenced by a Source or Output, you must first delete or update the Source or Output to no longer reference the VPC interface.

Creating a Request

data RemoveFlowVpcInterface

See: newRemoveFlowVpcInterface smart constructor.

Constructors

RemoveFlowVpcInterface'

Fields

flowArn :: Text

The flow that you want to remove a VPC interface from.

vpcInterfaceName :: Text

The name of the VPC interface that you want to remove.

newRemoveFlowVpcInterface

Arguments

:: Text
$sel:flowArn:RemoveFlowVpcInterface'

-> Text
$sel:vpcInterfaceName:RemoveFlowVpcInterface'

-> RemoveFlowVpcInterface

Create a value of RemoveFlowVpcInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flowArn:RemoveFlowVpcInterface', removeFlowVpcInterface_flowArn - The flow that you want to remove a VPC interface from.

$sel:vpcInterfaceName:RemoveFlowVpcInterface', removeFlowVpcInterface_vpcInterfaceName - The name of the VPC interface that you want to remove.

Request Lenses

removeFlowVpcInterface_flowArn :: Lens' RemoveFlowVpcInterface Text

The flow that you want to remove a VPC interface from.

removeFlowVpcInterface_vpcInterfaceName :: Lens' RemoveFlowVpcInterface Text

The name of the VPC interface that you want to remove.

Destructuring the Response

data RemoveFlowVpcInterfaceResponse

See: newRemoveFlowVpcInterfaceResponse smart constructor.

Constructors

RemoveFlowVpcInterfaceResponse'

Fields

flowArn :: Maybe Text

The ARN of the flow that is associated with the VPC interface you removed.

nonDeletedNetworkInterfaceIds :: Maybe [Text]

IDs of network interfaces associated with the removed VPC interface that Media Connect was unable to remove.

vpcInterfaceName :: Maybe Text

The name of the VPC interface that was removed.

httpStatus :: Int

The response's http status code.

newRemoveFlowVpcInterfaceResponse

Arguments

:: Int
$sel:httpStatus:RemoveFlowVpcInterfaceResponse'

-> RemoveFlowVpcInterfaceResponse

Create a value of RemoveFlowVpcInterfaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flowArn:RemoveFlowVpcInterface', removeFlowVpcInterfaceResponse_flowArn - The ARN of the flow that is associated with the VPC interface you removed.

$sel:nonDeletedNetworkInterfaceIds:RemoveFlowVpcInterfaceResponse', removeFlowVpcInterfaceResponse_nonDeletedNetworkInterfaceIds - IDs of network interfaces associated with the removed VPC interface that Media Connect was unable to remove.

$sel:vpcInterfaceName:RemoveFlowVpcInterface', removeFlowVpcInterfaceResponse_vpcInterfaceName - The name of the VPC interface that was removed.

$sel:httpStatus:RemoveFlowVpcInterfaceResponse', removeFlowVpcInterfaceResponse_httpStatus - The response's http status code.

Response Lenses

removeFlowVpcInterfaceResponse_flowArn :: Lens' RemoveFlowVpcInterfaceResponse (Maybe Text)

The ARN of the flow that is associated with the VPC interface you removed.

removeFlowVpcInterfaceResponse_nonDeletedNetworkInterfaceIds :: Lens' RemoveFlowVpcInterfaceResponse (Maybe [Text])

IDs of network interfaces associated with the removed VPC interface that Media Connect was unable to remove.

removeFlowVpcInterfaceResponse_vpcInterfaceName :: Lens' RemoveFlowVpcInterfaceResponse (Maybe Text)

The name of the VPC interface that was removed.

removeFlowVpcInterfaceResponse_httpStatus :: Lens' RemoveFlowVpcInterfaceResponse Int

The response's http status code.