Network.AWS.MediaConnect.Types.DestinationConfigurationRequest

Description

Documentation

data DestinationConfigurationRequest

The transport parameters that you want to associate with an outbound media stream.

See: newDestinationConfigurationRequest smart constructor.

Constructors

DestinationConfigurationRequest'

Fields

destinationIp :: Text

The IP address where you want MediaConnect to send contents of the media stream.

destinationPort :: Int

The port that you want MediaConnect to use when it distributes the media stream to the output.

interface :: InterfaceRequest

The VPC interface that you want to use for the media stream associated with the output.

newDestinationConfigurationRequest

Arguments

:: Text
$sel:destinationIp:DestinationConfigurationRequest'

-> Int
$sel:destinationPort:DestinationConfigurationRequest'

-> InterfaceRequest
$sel:interface:DestinationConfigurationRequest'

-> DestinationConfigurationRequest

Create a value of DestinationConfigurationRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destinationIp:DestinationConfigurationRequest', destinationConfigurationRequest_destinationIp - The IP address where you want MediaConnect to send contents of the media stream.

$sel:destinationPort:DestinationConfigurationRequest', destinationConfigurationRequest_destinationPort - The port that you want MediaConnect to use when it distributes the media stream to the output.

$sel:interface:DestinationConfigurationRequest', destinationConfigurationRequest_interface - The VPC interface that you want to use for the media stream associated with the output.

destinationConfigurationRequest_destinationIp :: Lens' DestinationConfigurationRequest Text

The IP address where you want MediaConnect to send contents of the media stream.

destinationConfigurationRequest_destinationPort :: Lens' DestinationConfigurationRequest Int

The port that you want MediaConnect to use when it distributes the media stream to the output.

destinationConfigurationRequest_interface :: Lens' DestinationConfigurationRequest InterfaceRequest

The VPC interface that you want to use for the media stream associated with the output.