Network.AWS.MediaConnect.Types.FailoverConfig

Description

Documentation

data FailoverConfig

The settings for source failover

See: newFailoverConfig smart constructor.

Constructors

FailoverConfig'

Fields

state :: Maybe State

recoveryWindow :: Maybe Int

Search window time to look for dash-7 packets

sourcePriority :: Maybe SourcePriority

The priority you want to assign to a source. You can have a primary stream and a backup stream or two equally prioritized streams.

failoverMode :: Maybe FailoverMode

The type of failover you choose for this flow. MERGE combines the source streams into a single stream, allowing graceful recovery from any single-source loss. FAILOVER allows switching between different streams.

newFailoverConfig :: FailoverConfig

Create a value of FailoverConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:FailoverConfig', failoverConfig_state - Undocumented member.

$sel:recoveryWindow:FailoverConfig', failoverConfig_recoveryWindow - Search window time to look for dash-7 packets

$sel:sourcePriority:FailoverConfig', failoverConfig_sourcePriority - The priority you want to assign to a source. You can have a primary stream and a backup stream or two equally prioritized streams.

$sel:failoverMode:FailoverConfig', failoverConfig_failoverMode - The type of failover you choose for this flow. MERGE combines the source streams into a single stream, allowing graceful recovery from any single-source loss. FAILOVER allows switching between different streams.

failoverConfig_state :: Lens' FailoverConfig (Maybe State)

Undocumented member.

failoverConfig_recoveryWindow :: Lens' FailoverConfig (Maybe Int)

Search window time to look for dash-7 packets

failoverConfig_sourcePriority :: Lens' FailoverConfig (Maybe SourcePriority)

The priority you want to assign to a source. You can have a primary stream and a backup stream or two equally prioritized streams.

failoverConfig_failoverMode :: Lens' FailoverConfig (Maybe FailoverMode)

The type of failover you choose for this flow. MERGE combines the source streams into a single stream, allowing graceful recovery from any single-source loss. FAILOVER allows switching between different streams.