Network.AWS.MediaConnect.Types.Fmtp

Description

Documentation

data Fmtp

FMTP

See: newFmtp smart constructor.

Constructors

Fmtp'

Fields

tcs :: Maybe Tcs

The transfer characteristic system (TCS) that is used in the video.

exactFramerate :: Maybe Text

The frame rate for the video stream, in frames/second. For example: 60000/1001. If you specify a whole number, MediaConnect uses a ratio of N/1. For example, if you specify 60, MediaConnect uses 60/1 as the exactFramerate.

par :: Maybe Text

The pixel aspect ratio (PAR) of the video.

scanMode :: Maybe ScanMode

The type of compression that was used to smooth the video’s appearance

range :: Maybe Range

The encoding range of the video.

channelOrder :: Maybe Text

The format of the audio channel.

colorimetry :: Maybe Colorimetry

The format that is used for the representation of color.

newFmtp :: Fmtp

Create a value of Fmtp with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tcs:Fmtp', fmtp_tcs - The transfer characteristic system (TCS) that is used in the video.

$sel:exactFramerate:Fmtp', fmtp_exactFramerate - The frame rate for the video stream, in frames/second. For example: 60000/1001. If you specify a whole number, MediaConnect uses a ratio of N/1. For example, if you specify 60, MediaConnect uses 60/1 as the exactFramerate.

$sel:par:Fmtp', fmtp_par - The pixel aspect ratio (PAR) of the video.

$sel:scanMode:Fmtp', fmtp_scanMode - The type of compression that was used to smooth the video’s appearance

$sel:range:Fmtp', fmtp_range - The encoding range of the video.

$sel:channelOrder:Fmtp', fmtp_channelOrder - The format of the audio channel.

$sel:colorimetry:Fmtp', fmtp_colorimetry - The format that is used for the representation of color.

fmtp_tcs :: Lens' Fmtp (Maybe Tcs)

The transfer characteristic system (TCS) that is used in the video.

fmtp_exactFramerate :: Lens' Fmtp (Maybe Text)

The frame rate for the video stream, in frames/second. For example: 60000/1001. If you specify a whole number, MediaConnect uses a ratio of N/1. For example, if you specify 60, MediaConnect uses 60/1 as the exactFramerate.

fmtp_par :: Lens' Fmtp (Maybe Text)

The pixel aspect ratio (PAR) of the video.

fmtp_scanMode :: Lens' Fmtp (Maybe ScanMode)

The type of compression that was used to smooth the video’s appearance

fmtp_range :: Lens' Fmtp (Maybe Range)

The encoding range of the video.

fmtp_channelOrder :: Lens' Fmtp (Maybe Text)

The format of the audio channel.

fmtp_colorimetry :: Lens' Fmtp (Maybe Colorimetry)

The format that is used for the representation of color.