Network.AWS.MediaConnect.Types.FmtpRequest

Description

Documentation

data FmtpRequest

The settings that you want to use to define the media stream.

See: newFmtpRequest smart constructor.

Constructors

FmtpRequest'

Fields

tcs :: Maybe Tcs

The transfer characteristic system (TCS) that is used in the video.

exactFramerate :: Maybe Text

The frame rate for the video stream, in frames/second. For example: 60000/1001. If you specify a whole number, MediaConnect uses a ratio of N/1. For example, if you specify 60, MediaConnect uses 60/1 as the exactFramerate.

par :: Maybe Text

The pixel aspect ratio (PAR) of the video.

scanMode :: Maybe ScanMode

The type of compression that was used to smooth the video’s appearance.

range :: Maybe Range

The encoding range of the video.

channelOrder :: Maybe Text

The format of the audio channel.

colorimetry :: Maybe Colorimetry

The format that is used for the representation of color.

newFmtpRequest :: FmtpRequest

Create a value of FmtpRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tcs:FmtpRequest', fmtpRequest_tcs - The transfer characteristic system (TCS) that is used in the video.

$sel:exactFramerate:FmtpRequest', fmtpRequest_exactFramerate - The frame rate for the video stream, in frames/second. For example: 60000/1001. If you specify a whole number, MediaConnect uses a ratio of N/1. For example, if you specify 60, MediaConnect uses 60/1 as the exactFramerate.

$sel:par:FmtpRequest', fmtpRequest_par - The pixel aspect ratio (PAR) of the video.

$sel:scanMode:FmtpRequest', fmtpRequest_scanMode - The type of compression that was used to smooth the video’s appearance.

$sel:range:FmtpRequest', fmtpRequest_range - The encoding range of the video.

$sel:channelOrder:FmtpRequest', fmtpRequest_channelOrder - The format of the audio channel.

$sel:colorimetry:FmtpRequest', fmtpRequest_colorimetry - The format that is used for the representation of color.

fmtpRequest_tcs :: Lens' FmtpRequest (Maybe Tcs)

The transfer characteristic system (TCS) that is used in the video.

fmtpRequest_exactFramerate :: Lens' FmtpRequest (Maybe Text)

The frame rate for the video stream, in frames/second. For example: 60000/1001. If you specify a whole number, MediaConnect uses a ratio of N/1. For example, if you specify 60, MediaConnect uses 60/1 as the exactFramerate.

fmtpRequest_par :: Lens' FmtpRequest (Maybe Text)

The pixel aspect ratio (PAR) of the video.

fmtpRequest_scanMode :: Lens' FmtpRequest (Maybe ScanMode)

The type of compression that was used to smooth the video’s appearance.

fmtpRequest_range :: Lens' FmtpRequest (Maybe Range)

The encoding range of the video.

fmtpRequest_channelOrder :: Lens' FmtpRequest (Maybe Text)

The format of the audio channel.

fmtpRequest_colorimetry :: Lens' FmtpRequest (Maybe Colorimetry)

The format that is used for the representation of color.