Network.AWS.MediaConnect.Types.ListedEntitlement

Description

Documentation

data ListedEntitlement

An entitlement that has been granted to you from other AWS accounts.

See: newListedEntitlement smart constructor.

Constructors

ListedEntitlement'

Fields

dataTransferSubscriberFeePercent :: Maybe Int

Percentage from 0-100 of the data transfer cost to be billed to the subscriber.

entitlementArn :: Text

The ARN of the entitlement.

entitlementName :: Text

The name of the entitlement.

newListedEntitlement

Arguments

:: Text
$sel:entitlementArn:ListedEntitlement'

-> Text
$sel:entitlementName:ListedEntitlement'

-> ListedEntitlement

Create a value of ListedEntitlement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataTransferSubscriberFeePercent:ListedEntitlement', listedEntitlement_dataTransferSubscriberFeePercent - Percentage from 0-100 of the data transfer cost to be billed to the subscriber.

$sel:entitlementArn:ListedEntitlement', listedEntitlement_entitlementArn - The ARN of the entitlement.

$sel:entitlementName:ListedEntitlement', listedEntitlement_entitlementName - The name of the entitlement.

listedEntitlement_dataTransferSubscriberFeePercent :: Lens' ListedEntitlement (Maybe Int)

Percentage from 0-100 of the data transfer cost to be billed to the subscriber.

listedEntitlement_entitlementArn :: Lens' ListedEntitlement Text

The ARN of the entitlement.

listedEntitlement_entitlementName :: Lens' ListedEntitlement Text

The name of the entitlement.