Network.AWS.MediaConnect.Types.MediaStream

Description

Documentation

data MediaStream

A single track or stream of media that contains video, audio, or ancillary data. After you add a media stream to a flow, you can associate it with sources and outputs on that flow, as long as they use the CDI protocol or the ST 2110 JPEG XS protocol. Each source or output can consist of one or many media streams.

See: newMediaStream smart constructor.

Constructors

MediaStream'

Fields

videoFormat :: Maybe Text

The resolution of the video.

attributes :: Maybe MediaStreamAttributes

Attributes that are related to the media stream.

clockRate :: Maybe Int

The sample rate for the stream. This value is measured in Hz.

description :: Maybe Text

A description that can help you quickly identify what your media stream is used for.

mediaStreamType :: MediaStreamType

The type of media stream.

mediaStreamId :: Int

A unique identifier for the media stream.

mediaStreamName :: Text

A name that helps you distinguish one media stream from another.

fmt :: Int

The format type number (sometimes referred to as RTP payload type) of the media stream. MediaConnect assigns this value to the media stream. For ST 2110 JPEG XS outputs, you need to provide this value to the receiver.

newMediaStream

Arguments

:: MediaStreamType
$sel:mediaStreamType:MediaStream'

-> Int
$sel:mediaStreamId:MediaStream'

-> Text
$sel:mediaStreamName:MediaStream'

-> Int
$sel:fmt:MediaStream'

-> MediaStream

Create a value of MediaStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:videoFormat:MediaStream', mediaStream_videoFormat - The resolution of the video.

$sel:attributes:MediaStream', mediaStream_attributes - Attributes that are related to the media stream.

$sel:clockRate:MediaStream', mediaStream_clockRate - The sample rate for the stream. This value is measured in Hz.

$sel:description:MediaStream', mediaStream_description - A description that can help you quickly identify what your media stream is used for.

$sel:mediaStreamType:MediaStream', mediaStream_mediaStreamType - The type of media stream.

$sel:mediaStreamId:MediaStream', mediaStream_mediaStreamId - A unique identifier for the media stream.

$sel:mediaStreamName:MediaStream', mediaStream_mediaStreamName - A name that helps you distinguish one media stream from another.

$sel:fmt:MediaStream', mediaStream_fmt - The format type number (sometimes referred to as RTP payload type) of the media stream. MediaConnect assigns this value to the media stream. For ST 2110 JPEG XS outputs, you need to provide this value to the receiver.

mediaStream_videoFormat :: Lens' MediaStream (Maybe Text)

The resolution of the video.

mediaStream_attributes :: Lens' MediaStream (Maybe MediaStreamAttributes)

Attributes that are related to the media stream.

mediaStream_clockRate :: Lens' MediaStream (Maybe Int)

The sample rate for the stream. This value is measured in Hz.

mediaStream_description :: Lens' MediaStream (Maybe Text)

A description that can help you quickly identify what your media stream is used for.

mediaStream_mediaStreamType :: Lens' MediaStream MediaStreamType

The type of media stream.

mediaStream_mediaStreamId :: Lens' MediaStream Int

A unique identifier for the media stream.

mediaStream_mediaStreamName :: Lens' MediaStream Text

A name that helps you distinguish one media stream from another.

mediaStream_fmt :: Lens' MediaStream Int

The format type number (sometimes referred to as RTP payload type) of the media stream. MediaConnect assigns this value to the media stream. For ST 2110 JPEG XS outputs, you need to provide this value to the receiver.