Network.AWS.MediaConnect.Types.Offering

Description

Documentation

data Offering

A savings plan that reserves a certain amount of outbound bandwidth usage at a discounted rate each month over a period of time.

See: newOffering smart constructor.

Constructors

Offering'

Fields

currencyCode :: Text

The type of currency that is used for billing. The currencyCode used for all reservations is US dollars.

offeringArn :: Text

The Amazon Resource Name (ARN) that MediaConnect assigns to the offering.

offeringDescription :: Text

A description of the offering.

durationUnits :: DurationUnits

The unit of measurement for the duration of the offering.

duration :: Int

The length of time that your reservation would be active.

pricePerUnit :: Text

The cost of a single unit. This value, in combination with priceUnits, makes up the rate.

resourceSpecification :: ResourceSpecification

A definition of the amount of outbound bandwidth that you would be reserving if you purchase the offering.

priceUnits :: PriceUnits

The unit of measurement that is used for billing. This value, in combination with pricePerUnit, makes up the rate.

newOffering

Arguments

:: Text
$sel:currencyCode:Offering'

-> Text
$sel:offeringArn:Offering'

-> Text
$sel:offeringDescription:Offering'

-> DurationUnits
$sel:durationUnits:Offering'

-> Int
$sel:duration:Offering'

-> Text
$sel:pricePerUnit:Offering'

-> ResourceSpecification
$sel:resourceSpecification:Offering'

-> PriceUnits
$sel:priceUnits:Offering'

-> Offering

Create a value of Offering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currencyCode:Offering', offering_currencyCode - The type of currency that is used for billing. The currencyCode used for all reservations is US dollars.

$sel:offeringArn:Offering', offering_offeringArn - The Amazon Resource Name (ARN) that MediaConnect assigns to the offering.

$sel:offeringDescription:Offering', offering_offeringDescription - A description of the offering.

$sel:durationUnits:Offering', offering_durationUnits - The unit of measurement for the duration of the offering.

$sel:duration:Offering', offering_duration - The length of time that your reservation would be active.

$sel:pricePerUnit:Offering', offering_pricePerUnit - The cost of a single unit. This value, in combination with priceUnits, makes up the rate.

$sel:resourceSpecification:Offering', offering_resourceSpecification - A definition of the amount of outbound bandwidth that you would be reserving if you purchase the offering.

$sel:priceUnits:Offering', offering_priceUnits - The unit of measurement that is used for billing. This value, in combination with pricePerUnit, makes up the rate.

offering_currencyCode :: Lens' Offering Text

The type of currency that is used for billing. The currencyCode used for all reservations is US dollars.

offering_offeringArn :: Lens' Offering Text

The Amazon Resource Name (ARN) that MediaConnect assigns to the offering.

offering_offeringDescription :: Lens' Offering Text

A description of the offering.

offering_durationUnits :: Lens' Offering DurationUnits

The unit of measurement for the duration of the offering.

offering_duration :: Lens' Offering Int

The length of time that your reservation would be active.

offering_pricePerUnit :: Lens' Offering Text

The cost of a single unit. This value, in combination with priceUnits, makes up the rate.

offering_resourceSpecification :: Lens' Offering ResourceSpecification

A definition of the amount of outbound bandwidth that you would be reserving if you purchase the offering.

offering_priceUnits :: Lens' Offering PriceUnits

The unit of measurement that is used for billing. This value, in combination with pricePerUnit, makes up the rate.