Network.AWS.MediaConnect.Types.Reservation

Description

Documentation

data Reservation

A pricing agreement for a discounted rate for a specific outbound bandwidth that your MediaConnect account will use each month over a specific time period. The discounted rate in the reservation applies to outbound bandwidth for all flows from your account until your account reaches the amount of bandwidth in your reservation. If you use more outbound bandwidth than the agreed upon amount in a single month, the overage is charged at the on-demand rate.

See: newReservation smart constructor.

Constructors

Reservation'

Fields

currencyCode :: Text

The type of currency that is used for billing. The currencyCode used for your reservation is US dollars.

reservationState :: ReservationState

The status of your reservation.

offeringArn :: Text

The Amazon Resource Name (ARN) that MediaConnect assigns to the offering.

reservationArn :: Text

The Amazon Resource Name (ARN) that MediaConnect assigns to the reservation when you purchase an offering.

start :: Text

The day and time that the reservation becomes active. You set this value when you purchase the offering.

offeringDescription :: Text

A description of the offering. MediaConnect defines this value in the offering.

reservationName :: Text

The name that you assigned to the reservation when you purchased the offering.

end :: Text

The day and time that this reservation expires. This value is calculated based on the start date and time that you set and the offering's duration.

duration :: Int

The length of time that this reservation is active. MediaConnect defines this value in the offering.

durationUnits :: DurationUnits

The unit of measurement for the duration of the reservation. MediaConnect defines this value in the offering.

pricePerUnit :: Text

The cost of a single unit. This value, in combination with priceUnits, makes up the rate. MediaConnect defines this value in the offering.

resourceSpecification :: ResourceSpecification

A definition of the amount of outbound bandwidth that you would be reserving if you purchase the offering. MediaConnect defines the values that make up the resourceSpecification in the offering.

priceUnits :: PriceUnits

The unit of measurement that is used for billing. This value, in combination with pricePerUnit, makes up the rate. MediaConnect defines this value in the offering.

newReservation

Arguments

:: Text
$sel:currencyCode:Reservation'

-> ReservationState
$sel:reservationState:Reservation'

-> Text
$sel:offeringArn:Reservation'

-> Text
$sel:reservationArn:Reservation'

-> Text
$sel:start:Reservation'

-> Text
$sel:offeringDescription:Reservation'

-> Text
$sel:reservationName:Reservation'

-> Text
$sel:end:Reservation'

-> Int
$sel:duration:Reservation'

-> DurationUnits
$sel:durationUnits:Reservation'

-> Text
$sel:pricePerUnit:Reservation'

-> ResourceSpecification
$sel:resourceSpecification:Reservation'

-> PriceUnits
$sel:priceUnits:Reservation'

-> Reservation

Create a value of Reservation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currencyCode:Reservation', reservation_currencyCode - The type of currency that is used for billing. The currencyCode used for your reservation is US dollars.

$sel:reservationState:Reservation', reservation_reservationState - The status of your reservation.

$sel:offeringArn:Reservation', reservation_offeringArn - The Amazon Resource Name (ARN) that MediaConnect assigns to the offering.

$sel:reservationArn:Reservation', reservation_reservationArn - The Amazon Resource Name (ARN) that MediaConnect assigns to the reservation when you purchase an offering.

$sel:start:Reservation', reservation_start - The day and time that the reservation becomes active. You set this value when you purchase the offering.

$sel:offeringDescription:Reservation', reservation_offeringDescription - A description of the offering. MediaConnect defines this value in the offering.

$sel:reservationName:Reservation', reservation_reservationName - The name that you assigned to the reservation when you purchased the offering.

$sel:end:Reservation', reservation_end - The day and time that this reservation expires. This value is calculated based on the start date and time that you set and the offering's duration.

$sel:duration:Reservation', reservation_duration - The length of time that this reservation is active. MediaConnect defines this value in the offering.

$sel:durationUnits:Reservation', reservation_durationUnits - The unit of measurement for the duration of the reservation. MediaConnect defines this value in the offering.

$sel:pricePerUnit:Reservation', reservation_pricePerUnit - The cost of a single unit. This value, in combination with priceUnits, makes up the rate. MediaConnect defines this value in the offering.

$sel:resourceSpecification:Reservation', reservation_resourceSpecification - A definition of the amount of outbound bandwidth that you would be reserving if you purchase the offering. MediaConnect defines the values that make up the resourceSpecification in the offering.

$sel:priceUnits:Reservation', reservation_priceUnits - The unit of measurement that is used for billing. This value, in combination with pricePerUnit, makes up the rate. MediaConnect defines this value in the offering.

reservation_currencyCode :: Lens' Reservation Text

The type of currency that is used for billing. The currencyCode used for your reservation is US dollars.

reservation_reservationState :: Lens' Reservation ReservationState

The status of your reservation.

reservation_offeringArn :: Lens' Reservation Text

The Amazon Resource Name (ARN) that MediaConnect assigns to the offering.

reservation_reservationArn :: Lens' Reservation Text

The Amazon Resource Name (ARN) that MediaConnect assigns to the reservation when you purchase an offering.

reservation_start :: Lens' Reservation Text

The day and time that the reservation becomes active. You set this value when you purchase the offering.

reservation_offeringDescription :: Lens' Reservation Text

A description of the offering. MediaConnect defines this value in the offering.

reservation_reservationName :: Lens' Reservation Text

The name that you assigned to the reservation when you purchased the offering.

reservation_end :: Lens' Reservation Text

The day and time that this reservation expires. This value is calculated based on the start date and time that you set and the offering's duration.

reservation_duration :: Lens' Reservation Int

The length of time that this reservation is active. MediaConnect defines this value in the offering.

reservation_durationUnits :: Lens' Reservation DurationUnits

The unit of measurement for the duration of the reservation. MediaConnect defines this value in the offering.

reservation_pricePerUnit :: Lens' Reservation Text

The cost of a single unit. This value, in combination with priceUnits, makes up the rate. MediaConnect defines this value in the offering.

reservation_resourceSpecification :: Lens' Reservation ResourceSpecification

A definition of the amount of outbound bandwidth that you would be reserving if you purchase the offering. MediaConnect defines the values that make up the resourceSpecification in the offering.

reservation_priceUnits :: Lens' Reservation PriceUnits

The unit of measurement that is used for billing. This value, in combination with pricePerUnit, makes up the rate. MediaConnect defines this value in the offering.