Network.AWS.MediaConnect.Types.Source

Description

Documentation

data Source

The settings for the source of the flow.

See: newSource smart constructor.

Constructors

Source'

Fields

entitlementArn :: Maybe Text

The ARN of the entitlement that allows you to subscribe to content that comes from another AWS account. The entitlement is set by the content originator and the ARN is generated as part of the originator's flow.

dataTransferSubscriberFeePercent :: Maybe Int

Percentage from 0-100 of the data transfer cost to be billed to the subscriber.

vpcInterfaceName :: Maybe Text

The name of the VPC interface that is used for this source.

decryption :: Maybe Encryption

The type of encryption that is used on the content ingested from this source.

ingestIp :: Maybe Text

The IP address that the flow will be listening on for incoming content.

ingestPort :: Maybe Int

The port that the flow will be listening on for incoming content.

transport :: Maybe Transport

Attributes related to the transport stream that are used in the source.

description :: Maybe Text

A description for the source. This value is not used or seen outside of the current AWS Elemental MediaConnect account.

mediaStreamSourceConfigurations :: Maybe [MediaStreamSourceConfiguration]

The media streams that are associated with the source, and the parameters for those associations.

whitelistCidr :: Maybe Text

The range of IP addresses that should be allowed to contribute content to your source. These IP addresses should be in the form of a Classless Inter-Domain Routing (CIDR) block; for example, 10.0.0.0/16.

sourceArn :: Text

The ARN of the source.

name :: Text

The name of the source.

newSource

Arguments

:: Text
$sel:sourceArn:Source'

-> Text
$sel:name:Source'

-> Source

Create a value of Source with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entitlementArn:Source', source_entitlementArn - The ARN of the entitlement that allows you to subscribe to content that comes from another AWS account. The entitlement is set by the content originator and the ARN is generated as part of the originator's flow.

$sel:dataTransferSubscriberFeePercent:Source', source_dataTransferSubscriberFeePercent - Percentage from 0-100 of the data transfer cost to be billed to the subscriber.

$sel:vpcInterfaceName:Source', source_vpcInterfaceName - The name of the VPC interface that is used for this source.

$sel:decryption:Source', source_decryption - The type of encryption that is used on the content ingested from this source.

$sel:ingestIp:Source', source_ingestIp - The IP address that the flow will be listening on for incoming content.

$sel:ingestPort:Source', source_ingestPort - The port that the flow will be listening on for incoming content.

$sel:transport:Source', source_transport - Attributes related to the transport stream that are used in the source.

$sel:description:Source', source_description - A description for the source. This value is not used or seen outside of the current AWS Elemental MediaConnect account.

$sel:mediaStreamSourceConfigurations:Source', source_mediaStreamSourceConfigurations - The media streams that are associated with the source, and the parameters for those associations.

$sel:whitelistCidr:Source', source_whitelistCidr - The range of IP addresses that should be allowed to contribute content to your source. These IP addresses should be in the form of a Classless Inter-Domain Routing (CIDR) block; for example, 10.0.0.0/16.

$sel:sourceArn:Source', source_sourceArn - The ARN of the source.

$sel:name:Source', source_name - The name of the source.

source_entitlementArn :: Lens' Source (Maybe Text)

The ARN of the entitlement that allows you to subscribe to content that comes from another AWS account. The entitlement is set by the content originator and the ARN is generated as part of the originator's flow.

source_dataTransferSubscriberFeePercent :: Lens' Source (Maybe Int)

Percentage from 0-100 of the data transfer cost to be billed to the subscriber.

source_vpcInterfaceName :: Lens' Source (Maybe Text)

The name of the VPC interface that is used for this source.

source_decryption :: Lens' Source (Maybe Encryption)

The type of encryption that is used on the content ingested from this source.

source_ingestIp :: Lens' Source (Maybe Text)

The IP address that the flow will be listening on for incoming content.

source_ingestPort :: Lens' Source (Maybe Int)

The port that the flow will be listening on for incoming content.

source_transport :: Lens' Source (Maybe Transport)

Attributes related to the transport stream that are used in the source.

source_description :: Lens' Source (Maybe Text)

A description for the source. This value is not used or seen outside of the current AWS Elemental MediaConnect account.

source_mediaStreamSourceConfigurations :: Lens' Source (Maybe [MediaStreamSourceConfiguration])

The media streams that are associated with the source, and the parameters for those associations.

source_whitelistCidr :: Lens' Source (Maybe Text)

The range of IP addresses that should be allowed to contribute content to your source. These IP addresses should be in the form of a Classless Inter-Domain Routing (CIDR) block; for example, 10.0.0.0/16.

source_sourceArn :: Lens' Source Text

The ARN of the source.

source_name :: Lens' Source Text

The name of the source.