Network.AWS.MediaConnect.Types.VpcInterface

Description

Documentation

data VpcInterface

The settings for a VPC Source.

See: newVpcInterface smart constructor.

Constructors

VpcInterface'

Fields

networkInterfaceType :: NetworkInterfaceType

The type of network interface.

networkInterfaceIds :: [Text]

IDs of the network interfaces created in customer's account by MediaConnect.

subnetId :: Text

Subnet must be in the AZ of the Flow

securityGroupIds :: [Text]

Security Group IDs to be used on ENI.

roleArn :: Text

Role Arn MediaConnect can assumes to create ENIs in customer's account

name :: Text

Immutable and has to be a unique against other VpcInterfaces in this Flow

newVpcInterface

Arguments

:: NetworkInterfaceType
$sel:networkInterfaceType:VpcInterface'

-> Text
$sel:subnetId:VpcInterface'

-> Text
$sel:roleArn:VpcInterface'

-> Text
$sel:name:VpcInterface'

-> VpcInterface

Create a value of VpcInterface with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkInterfaceType:VpcInterface', vpcInterface_networkInterfaceType - The type of network interface.

$sel:networkInterfaceIds:VpcInterface', vpcInterface_networkInterfaceIds - IDs of the network interfaces created in customer's account by MediaConnect.

$sel:subnetId:VpcInterface', vpcInterface_subnetId - Subnet must be in the AZ of the Flow

$sel:securityGroupIds:VpcInterface', vpcInterface_securityGroupIds - Security Group IDs to be used on ENI.

$sel:roleArn:VpcInterface', vpcInterface_roleArn - Role Arn MediaConnect can assumes to create ENIs in customer's account

$sel:name:VpcInterface', vpcInterface_name - Immutable and has to be a unique against other VpcInterfaces in this Flow

vpcInterface_networkInterfaceType :: Lens' VpcInterface NetworkInterfaceType

The type of network interface.

vpcInterface_networkInterfaceIds :: Lens' VpcInterface [Text]

IDs of the network interfaces created in customer's account by MediaConnect.

vpcInterface_subnetId :: Lens' VpcInterface Text

Subnet must be in the AZ of the Flow

vpcInterface_securityGroupIds :: Lens' VpcInterface [Text]

Security Group IDs to be used on ENI.

vpcInterface_roleArn :: Lens' VpcInterface Text

Role Arn MediaConnect can assumes to create ENIs in customer's account

vpcInterface_name :: Lens' VpcInterface Text

Immutable and has to be a unique against other VpcInterfaces in this Flow