Network.AWS.MediaConnect.UpdateFlowMediaStream

Description

Updates an existing media stream.

Creating a Request

data UpdateFlowMediaStream

The fields that you want to update in the media stream.

See: newUpdateFlowMediaStream smart constructor.

Constructors

UpdateFlowMediaStream'

Fields

videoFormat :: Maybe Text

The resolution of the video.

mediaStreamType :: Maybe MediaStreamType

The type of media stream.

attributes :: Maybe MediaStreamAttributesRequest

The attributes that you want to assign to the media stream.

clockRate :: Maybe Int

The sample rate (in Hz) for the stream. If the media stream type is video or ancillary data, set this value to 90000. If the media stream type is audio, set this value to either 48000 or 96000.

description :: Maybe Text

Description

flowArn :: Text

The Amazon Resource Name (ARN) of the flow.

mediaStreamName :: Text

The name of the media stream that you want to update.

newUpdateFlowMediaStream

Arguments

:: Text
$sel:flowArn:UpdateFlowMediaStream'

-> Text
$sel:mediaStreamName:UpdateFlowMediaStream'

-> UpdateFlowMediaStream

Create a value of UpdateFlowMediaStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:videoFormat:UpdateFlowMediaStream', updateFlowMediaStream_videoFormat - The resolution of the video.

$sel:mediaStreamType:UpdateFlowMediaStream', updateFlowMediaStream_mediaStreamType - The type of media stream.

$sel:attributes:UpdateFlowMediaStream', updateFlowMediaStream_attributes - The attributes that you want to assign to the media stream.

$sel:clockRate:UpdateFlowMediaStream', updateFlowMediaStream_clockRate - The sample rate (in Hz) for the stream. If the media stream type is video or ancillary data, set this value to 90000. If the media stream type is audio, set this value to either 48000 or 96000.

$sel:description:UpdateFlowMediaStream', updateFlowMediaStream_description - Description

$sel:flowArn:UpdateFlowMediaStream', updateFlowMediaStream_flowArn - The Amazon Resource Name (ARN) of the flow.

$sel:mediaStreamName:UpdateFlowMediaStream', updateFlowMediaStream_mediaStreamName - The name of the media stream that you want to update.

Request Lenses

updateFlowMediaStream_videoFormat :: Lens' UpdateFlowMediaStream (Maybe Text)

The resolution of the video.

updateFlowMediaStream_mediaStreamType :: Lens' UpdateFlowMediaStream (Maybe MediaStreamType)

The type of media stream.

updateFlowMediaStream_attributes :: Lens' UpdateFlowMediaStream (Maybe MediaStreamAttributesRequest)

The attributes that you want to assign to the media stream.

updateFlowMediaStream_clockRate :: Lens' UpdateFlowMediaStream (Maybe Int)

The sample rate (in Hz) for the stream. If the media stream type is video or ancillary data, set this value to 90000. If the media stream type is audio, set this value to either 48000 or 96000.

updateFlowMediaStream_description :: Lens' UpdateFlowMediaStream (Maybe Text)

Description

updateFlowMediaStream_flowArn :: Lens' UpdateFlowMediaStream Text

The Amazon Resource Name (ARN) of the flow.

updateFlowMediaStream_mediaStreamName :: Lens' UpdateFlowMediaStream Text

The name of the media stream that you want to update.

Destructuring the Response

data UpdateFlowMediaStreamResponse

See: newUpdateFlowMediaStreamResponse smart constructor.

Constructors

UpdateFlowMediaStreamResponse'

Fields

flowArn :: Maybe Text

The ARN of the flow that is associated with the media stream that you updated.

mediaStream :: Maybe MediaStream

The media stream that you updated.

httpStatus :: Int

The response's http status code.

newUpdateFlowMediaStreamResponse

Arguments

:: Int
$sel:httpStatus:UpdateFlowMediaStreamResponse'

-> UpdateFlowMediaStreamResponse

Create a value of UpdateFlowMediaStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flowArn:UpdateFlowMediaStream', updateFlowMediaStreamResponse_flowArn - The ARN of the flow that is associated with the media stream that you updated.

$sel:mediaStream:UpdateFlowMediaStreamResponse', updateFlowMediaStreamResponse_mediaStream - The media stream that you updated.

$sel:httpStatus:UpdateFlowMediaStreamResponse', updateFlowMediaStreamResponse_httpStatus - The response's http status code.

Response Lenses

updateFlowMediaStreamResponse_flowArn :: Lens' UpdateFlowMediaStreamResponse (Maybe Text)

The ARN of the flow that is associated with the media stream that you updated.

updateFlowMediaStreamResponse_mediaStream :: Lens' UpdateFlowMediaStreamResponse (Maybe MediaStream)

The media stream that you updated.

updateFlowMediaStreamResponse_httpStatus :: Lens' UpdateFlowMediaStreamResponse Int

The response's http status code.