Network.AWS.MediaConnect.UpdateFlowOutput

Description

Updates an existing flow output.

Creating a Request

data UpdateFlowOutput

The fields that you want to update in the output.

See: newUpdateFlowOutput smart constructor.

Constructors

UpdateFlowOutput'

Fields

destination :: Maybe Text

The IP address where you want to send the output.

maxLatency :: Maybe Int

The maximum latency in milliseconds for Zixi-based streams.

mediaStreamOutputConfigurations :: Maybe [MediaStreamOutputConfigurationRequest]

The media streams that are associated with the output, and the parameters for those associations.

protocol :: Maybe Protocol

The protocol to use for the output.

encryption :: Maybe UpdateEncryption

The type of key used for the encryption. If no keyType is provided, the service will use the default setting (static-key).

cidrAllowList :: Maybe [Text]

The range of IP addresses that should be allowed to initiate output requests to this flow. These IP addresses should be in the form of a Classless Inter-Domain Routing (CIDR) block; for example, 10.0.0.0/16.

smoothingLatency :: Maybe Int

The smoothing latency in milliseconds for RIST, RTP, and RTP-FEC streams.

minLatency :: Maybe Int

The minimum latency in milliseconds for SRT-based streams. In streams that use the SRT protocol, this value that you set on your MediaConnect source or output represents the minimal potential latency of that connection. The latency of the stream is set to the highest number between the sender’s minimum latency and the receiver’s minimum latency.

description :: Maybe Text

A description of the output. This description appears only on the AWS Elemental MediaConnect console and will not be seen by the end user.

port :: Maybe Int

The port to use when content is distributed to this output.

streamId :: Maybe Text

The stream ID that you want to use for this transport. This parameter applies only to Zixi-based streams.

remoteId :: Maybe Text

The remote ID for the Zixi-pull stream.

vpcInterfaceAttachment :: Maybe VpcInterfaceAttachment

The name of the VPC interface attachment to use for this output.

flowArn :: Text

The flow that is associated with the output that you want to update.

outputArn :: Text

The ARN of the output that you want to update.

newUpdateFlowOutput

Arguments

:: Text
$sel:flowArn:UpdateFlowOutput'

-> Text
$sel:outputArn:UpdateFlowOutput'

-> UpdateFlowOutput

Create a value of UpdateFlowOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:destination:UpdateFlowOutput', updateFlowOutput_destination - The IP address where you want to send the output.

$sel:maxLatency:UpdateFlowOutput', updateFlowOutput_maxLatency - The maximum latency in milliseconds for Zixi-based streams.

$sel:mediaStreamOutputConfigurations:UpdateFlowOutput', updateFlowOutput_mediaStreamOutputConfigurations - The media streams that are associated with the output, and the parameters for those associations.

$sel:protocol:UpdateFlowOutput', updateFlowOutput_protocol - The protocol to use for the output.

$sel:encryption:UpdateFlowOutput', updateFlowOutput_encryption - The type of key used for the encryption. If no keyType is provided, the service will use the default setting (static-key).

$sel:cidrAllowList:UpdateFlowOutput', updateFlowOutput_cidrAllowList - The range of IP addresses that should be allowed to initiate output requests to this flow. These IP addresses should be in the form of a Classless Inter-Domain Routing (CIDR) block; for example, 10.0.0.0/16.

$sel:smoothingLatency:UpdateFlowOutput', updateFlowOutput_smoothingLatency - The smoothing latency in milliseconds for RIST, RTP, and RTP-FEC streams.

$sel:minLatency:UpdateFlowOutput', updateFlowOutput_minLatency - The minimum latency in milliseconds for SRT-based streams. In streams that use the SRT protocol, this value that you set on your MediaConnect source or output represents the minimal potential latency of that connection. The latency of the stream is set to the highest number between the sender’s minimum latency and the receiver’s minimum latency.

$sel:description:UpdateFlowOutput', updateFlowOutput_description - A description of the output. This description appears only on the AWS Elemental MediaConnect console and will not be seen by the end user.

$sel:port:UpdateFlowOutput', updateFlowOutput_port - The port to use when content is distributed to this output.

$sel:streamId:UpdateFlowOutput', updateFlowOutput_streamId - The stream ID that you want to use for this transport. This parameter applies only to Zixi-based streams.

$sel:remoteId:UpdateFlowOutput', updateFlowOutput_remoteId - The remote ID for the Zixi-pull stream.

$sel:vpcInterfaceAttachment:UpdateFlowOutput', updateFlowOutput_vpcInterfaceAttachment - The name of the VPC interface attachment to use for this output.

$sel:flowArn:UpdateFlowOutput', updateFlowOutput_flowArn - The flow that is associated with the output that you want to update.

$sel:outputArn:UpdateFlowOutput', updateFlowOutput_outputArn - The ARN of the output that you want to update.

Request Lenses

updateFlowOutput_destination :: Lens' UpdateFlowOutput (Maybe Text)

The IP address where you want to send the output.

updateFlowOutput_maxLatency :: Lens' UpdateFlowOutput (Maybe Int)

The maximum latency in milliseconds for Zixi-based streams.

updateFlowOutput_mediaStreamOutputConfigurations :: Lens' UpdateFlowOutput (Maybe [MediaStreamOutputConfigurationRequest])

The media streams that are associated with the output, and the parameters for those associations.

updateFlowOutput_protocol :: Lens' UpdateFlowOutput (Maybe Protocol)

The protocol to use for the output.

updateFlowOutput_encryption :: Lens' UpdateFlowOutput (Maybe UpdateEncryption)

The type of key used for the encryption. If no keyType is provided, the service will use the default setting (static-key).

updateFlowOutput_cidrAllowList :: Lens' UpdateFlowOutput (Maybe [Text])

The range of IP addresses that should be allowed to initiate output requests to this flow. These IP addresses should be in the form of a Classless Inter-Domain Routing (CIDR) block; for example, 10.0.0.0/16.

updateFlowOutput_smoothingLatency :: Lens' UpdateFlowOutput (Maybe Int)

The smoothing latency in milliseconds for RIST, RTP, and RTP-FEC streams.

updateFlowOutput_minLatency :: Lens' UpdateFlowOutput (Maybe Int)

The minimum latency in milliseconds for SRT-based streams. In streams that use the SRT protocol, this value that you set on your MediaConnect source or output represents the minimal potential latency of that connection. The latency of the stream is set to the highest number between the sender’s minimum latency and the receiver’s minimum latency.

updateFlowOutput_description :: Lens' UpdateFlowOutput (Maybe Text)

A description of the output. This description appears only on the AWS Elemental MediaConnect console and will not be seen by the end user.

updateFlowOutput_port :: Lens' UpdateFlowOutput (Maybe Int)

The port to use when content is distributed to this output.

updateFlowOutput_streamId :: Lens' UpdateFlowOutput (Maybe Text)

The stream ID that you want to use for this transport. This parameter applies only to Zixi-based streams.

updateFlowOutput_remoteId :: Lens' UpdateFlowOutput (Maybe Text)

The remote ID for the Zixi-pull stream.

updateFlowOutput_vpcInterfaceAttachment :: Lens' UpdateFlowOutput (Maybe VpcInterfaceAttachment)

The name of the VPC interface attachment to use for this output.

updateFlowOutput_flowArn :: Lens' UpdateFlowOutput Text

The flow that is associated with the output that you want to update.

updateFlowOutput_outputArn :: Lens' UpdateFlowOutput Text

The ARN of the output that you want to update.

Destructuring the Response

data UpdateFlowOutputResponse

See: newUpdateFlowOutputResponse smart constructor.

Constructors

UpdateFlowOutputResponse'

Fields

flowArn :: Maybe Text

The ARN of the flow that is associated with the updated output.

output :: Maybe Output

The new settings of the output that you updated.

httpStatus :: Int

The response's http status code.

newUpdateFlowOutputResponse

Arguments

:: Int
$sel:httpStatus:UpdateFlowOutputResponse'

-> UpdateFlowOutputResponse

Create a value of UpdateFlowOutputResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flowArn:UpdateFlowOutput', updateFlowOutputResponse_flowArn - The ARN of the flow that is associated with the updated output.

$sel:output:UpdateFlowOutputResponse', updateFlowOutputResponse_output - The new settings of the output that you updated.

$sel:httpStatus:UpdateFlowOutputResponse', updateFlowOutputResponse_httpStatus - The response's http status code.

Response Lenses

updateFlowOutputResponse_flowArn :: Lens' UpdateFlowOutputResponse (Maybe Text)

The ARN of the flow that is associated with the updated output.

updateFlowOutputResponse_output :: Lens' UpdateFlowOutputResponse (Maybe Output)

The new settings of the output that you updated.

updateFlowOutputResponse_httpStatus :: Lens' UpdateFlowOutputResponse Int

The response's http status code.