Network.AWS.MediaConnect.UpdateFlowSource

Description

Updates the source of a flow.

Creating a Request

data UpdateFlowSource

A request to update the source of a flow.

See: newUpdateFlowSource smart constructor.

Constructors

UpdateFlowSource'

Fields

entitlementArn :: Maybe Text

The ARN of the entitlement that allows you to subscribe to this flow. The entitlement is set by the flow originator, and the ARN is generated as part of the originator's flow.

maxLatency :: Maybe Int

The maximum latency in milliseconds. This parameter applies only to RIST-based and Zixi-based streams.

vpcInterfaceName :: Maybe Text

The name of the VPC interface to use for this source.

decryption :: Maybe UpdateEncryption

The type of encryption used on the content ingested from this source.

maxSyncBuffer :: Maybe Int

The size of the buffer (in milliseconds) to use to sync incoming source data.

protocol :: Maybe Protocol

The protocol that is used by the source.

minLatency :: Maybe Int

The minimum latency in milliseconds for SRT-based streams. In streams that use the SRT protocol, this value that you set on your MediaConnect source or output represents the minimal potential latency of that connection. The latency of the stream is set to the highest number between the sender’s minimum latency and the receiver’s minimum latency.

ingestPort :: Maybe Int

The port that the flow will be listening on for incoming content.

description :: Maybe Text

A description for the source. This value is not used or seen outside of the current AWS Elemental MediaConnect account.

mediaStreamSourceConfigurations :: Maybe [MediaStreamSourceConfigurationRequest]

The media streams that are associated with the source, and the parameters for those associations.

whitelistCidr :: Maybe Text

The range of IP addresses that should be allowed to contribute content to your source. These IP addresses should be in the form of a Classless Inter-Domain Routing (CIDR) block; for example, 10.0.0.0/16.

maxBitrate :: Maybe Int

The smoothing max bitrate for RIST, RTP, and RTP-FEC streams.

streamId :: Maybe Text

The stream ID that you want to use for this transport. This parameter applies only to Zixi-based streams.

flowArn :: Text

The flow that is associated with the source that you want to update.

sourceArn :: Text

The ARN of the source that you want to update.

newUpdateFlowSource

Arguments

:: Text
$sel:flowArn:UpdateFlowSource'

-> Text
$sel:sourceArn:UpdateFlowSource'

-> UpdateFlowSource

Create a value of UpdateFlowSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entitlementArn:UpdateFlowSource', updateFlowSource_entitlementArn - The ARN of the entitlement that allows you to subscribe to this flow. The entitlement is set by the flow originator, and the ARN is generated as part of the originator's flow.

$sel:maxLatency:UpdateFlowSource', updateFlowSource_maxLatency - The maximum latency in milliseconds. This parameter applies only to RIST-based and Zixi-based streams.

$sel:vpcInterfaceName:UpdateFlowSource', updateFlowSource_vpcInterfaceName - The name of the VPC interface to use for this source.

$sel:decryption:UpdateFlowSource', updateFlowSource_decryption - The type of encryption used on the content ingested from this source.

$sel:maxSyncBuffer:UpdateFlowSource', updateFlowSource_maxSyncBuffer - The size of the buffer (in milliseconds) to use to sync incoming source data.

$sel:protocol:UpdateFlowSource', updateFlowSource_protocol - The protocol that is used by the source.

$sel:minLatency:UpdateFlowSource', updateFlowSource_minLatency - The minimum latency in milliseconds for SRT-based streams. In streams that use the SRT protocol, this value that you set on your MediaConnect source or output represents the minimal potential latency of that connection. The latency of the stream is set to the highest number between the sender’s minimum latency and the receiver’s minimum latency.

$sel:ingestPort:UpdateFlowSource', updateFlowSource_ingestPort - The port that the flow will be listening on for incoming content.

$sel:description:UpdateFlowSource', updateFlowSource_description - A description for the source. This value is not used or seen outside of the current AWS Elemental MediaConnect account.

$sel:mediaStreamSourceConfigurations:UpdateFlowSource', updateFlowSource_mediaStreamSourceConfigurations - The media streams that are associated with the source, and the parameters for those associations.

$sel:whitelistCidr:UpdateFlowSource', updateFlowSource_whitelistCidr - The range of IP addresses that should be allowed to contribute content to your source. These IP addresses should be in the form of a Classless Inter-Domain Routing (CIDR) block; for example, 10.0.0.0/16.

$sel:maxBitrate:UpdateFlowSource', updateFlowSource_maxBitrate - The smoothing max bitrate for RIST, RTP, and RTP-FEC streams.

$sel:streamId:UpdateFlowSource', updateFlowSource_streamId - The stream ID that you want to use for this transport. This parameter applies only to Zixi-based streams.

$sel:flowArn:UpdateFlowSource', updateFlowSource_flowArn - The flow that is associated with the source that you want to update.

$sel:sourceArn:UpdateFlowSource', updateFlowSource_sourceArn - The ARN of the source that you want to update.

Request Lenses

updateFlowSource_entitlementArn :: Lens' UpdateFlowSource (Maybe Text)

The ARN of the entitlement that allows you to subscribe to this flow. The entitlement is set by the flow originator, and the ARN is generated as part of the originator's flow.

updateFlowSource_maxLatency :: Lens' UpdateFlowSource (Maybe Int)

The maximum latency in milliseconds. This parameter applies only to RIST-based and Zixi-based streams.

updateFlowSource_vpcInterfaceName :: Lens' UpdateFlowSource (Maybe Text)

The name of the VPC interface to use for this source.

updateFlowSource_decryption :: Lens' UpdateFlowSource (Maybe UpdateEncryption)

The type of encryption used on the content ingested from this source.

updateFlowSource_maxSyncBuffer :: Lens' UpdateFlowSource (Maybe Int)

The size of the buffer (in milliseconds) to use to sync incoming source data.

updateFlowSource_protocol :: Lens' UpdateFlowSource (Maybe Protocol)

The protocol that is used by the source.

updateFlowSource_minLatency :: Lens' UpdateFlowSource (Maybe Int)

The minimum latency in milliseconds for SRT-based streams. In streams that use the SRT protocol, this value that you set on your MediaConnect source or output represents the minimal potential latency of that connection. The latency of the stream is set to the highest number between the sender’s minimum latency and the receiver’s minimum latency.

updateFlowSource_ingestPort :: Lens' UpdateFlowSource (Maybe Int)

The port that the flow will be listening on for incoming content.

updateFlowSource_description :: Lens' UpdateFlowSource (Maybe Text)

A description for the source. This value is not used or seen outside of the current AWS Elemental MediaConnect account.

updateFlowSource_mediaStreamSourceConfigurations :: Lens' UpdateFlowSource (Maybe [MediaStreamSourceConfigurationRequest])

The media streams that are associated with the source, and the parameters for those associations.

updateFlowSource_whitelistCidr :: Lens' UpdateFlowSource (Maybe Text)

The range of IP addresses that should be allowed to contribute content to your source. These IP addresses should be in the form of a Classless Inter-Domain Routing (CIDR) block; for example, 10.0.0.0/16.

updateFlowSource_maxBitrate :: Lens' UpdateFlowSource (Maybe Int)

The smoothing max bitrate for RIST, RTP, and RTP-FEC streams.

updateFlowSource_streamId :: Lens' UpdateFlowSource (Maybe Text)

The stream ID that you want to use for this transport. This parameter applies only to Zixi-based streams.

updateFlowSource_flowArn :: Lens' UpdateFlowSource Text

The flow that is associated with the source that you want to update.

updateFlowSource_sourceArn :: Lens' UpdateFlowSource Text

The ARN of the source that you want to update.

Destructuring the Response

data UpdateFlowSourceResponse

See: newUpdateFlowSourceResponse smart constructor.

Constructors

UpdateFlowSourceResponse'

Fields

flowArn :: Maybe Text

The ARN of the flow that you want to update.

source :: Maybe Source

The settings for the source of the flow.

httpStatus :: Int

The response's http status code.

newUpdateFlowSourceResponse

Arguments

:: Int
$sel:httpStatus:UpdateFlowSourceResponse'

-> UpdateFlowSourceResponse

Create a value of UpdateFlowSourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:flowArn:UpdateFlowSource', updateFlowSourceResponse_flowArn - The ARN of the flow that you want to update.

$sel:source:UpdateFlowSourceResponse', updateFlowSourceResponse_source - The settings for the source of the flow.

$sel:httpStatus:UpdateFlowSourceResponse', updateFlowSourceResponse_httpStatus - The response's http status code.

Response Lenses

updateFlowSourceResponse_flowArn :: Lens' UpdateFlowSourceResponse (Maybe Text)

The ARN of the flow that you want to update.

updateFlowSourceResponse_source :: Lens' UpdateFlowSourceResponse (Maybe Source)

The settings for the source of the flow.

updateFlowSourceResponse_httpStatus :: Lens' UpdateFlowSourceResponse Int

The response's http status code.