Network.AWS.MediaConvert.CreateJobTemplate

Description

Create a new job template. For information about job templates see the User Guide at http://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html

Creating a Request

data CreateJobTemplate

See: newCreateJobTemplate smart constructor.

Constructors

CreateJobTemplate'

Fields

accelerationSettings :: Maybe AccelerationSettings

Accelerated transcoding can significantly speed up jobs with long, visually complex content. Outputs that use this feature incur pro-tier pricing. For information about feature limitations, see the AWS Elemental MediaConvert User Guide.

category :: Maybe Text

Optional. A category for the job template you are creating

statusUpdateInterval :: Maybe StatusUpdateInterval

Specify how often MediaConvert sends STATUS_UPDATE events to Amazon CloudWatch Events. Set the interval, in seconds, between status updates. MediaConvert sends an update at this interval from the time the service begins processing your job to the time it completes the transcode or encounters an error.

priority :: Maybe Int

Specify the relative priority for this job. In any given queue, the service begins processing the job with the highest value first. When more than one job has the same priority, the service begins processing the job that you submitted first. If you don't specify a priority, the service uses the default value 0.

queue :: Maybe Text

Optional. The queue that jobs created from this template are assigned to. If you don't specify this, jobs will go to the default queue.

tags :: Maybe (HashMap Text Text)

The tags that you want to add to the resource. You can tag resources with a key-value pair or with only a key.

description :: Maybe Text

Optional. A description of the job template you are creating.

hopDestinations :: Maybe [HopDestination]

Optional. Use queue hopping to avoid overly long waits in the backlog of the queue that you submit your job to. Specify an alternate queue and the maximum time that your job will wait in the initial queue before hopping. For more information about this feature, see the AWS Elemental MediaConvert User Guide.

settings :: JobTemplateSettings

JobTemplateSettings contains all the transcode settings saved in the template that will be applied to jobs created from it.

name :: Text

The name of the job template you are creating.

newCreateJobTemplate

Arguments

:: JobTemplateSettings
$sel:settings:CreateJobTemplate'

-> Text
$sel:name:CreateJobTemplate'

-> CreateJobTemplate

Create a value of CreateJobTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accelerationSettings:CreateJobTemplate', createJobTemplate_accelerationSettings - Accelerated transcoding can significantly speed up jobs with long, visually complex content. Outputs that use this feature incur pro-tier pricing. For information about feature limitations, see the AWS Elemental MediaConvert User Guide.

$sel:category:CreateJobTemplate', createJobTemplate_category - Optional. A category for the job template you are creating

$sel:statusUpdateInterval:CreateJobTemplate', createJobTemplate_statusUpdateInterval - Specify how often MediaConvert sends STATUS_UPDATE events to Amazon CloudWatch Events. Set the interval, in seconds, between status updates. MediaConvert sends an update at this interval from the time the service begins processing your job to the time it completes the transcode or encounters an error.

$sel:priority:CreateJobTemplate', createJobTemplate_priority - Specify the relative priority for this job. In any given queue, the service begins processing the job with the highest value first. When more than one job has the same priority, the service begins processing the job that you submitted first. If you don't specify a priority, the service uses the default value 0.

$sel:queue:CreateJobTemplate', createJobTemplate_queue - Optional. The queue that jobs created from this template are assigned to. If you don't specify this, jobs will go to the default queue.

$sel:tags:CreateJobTemplate', createJobTemplate_tags - The tags that you want to add to the resource. You can tag resources with a key-value pair or with only a key.

$sel:description:CreateJobTemplate', createJobTemplate_description - Optional. A description of the job template you are creating.

$sel:hopDestinations:CreateJobTemplate', createJobTemplate_hopDestinations - Optional. Use queue hopping to avoid overly long waits in the backlog of the queue that you submit your job to. Specify an alternate queue and the maximum time that your job will wait in the initial queue before hopping. For more information about this feature, see the AWS Elemental MediaConvert User Guide.

$sel:settings:CreateJobTemplate', createJobTemplate_settings - JobTemplateSettings contains all the transcode settings saved in the template that will be applied to jobs created from it.

$sel:name:CreateJobTemplate', createJobTemplate_name - The name of the job template you are creating.

Request Lenses

createJobTemplate_accelerationSettings :: Lens' CreateJobTemplate (Maybe AccelerationSettings)

Accelerated transcoding can significantly speed up jobs with long, visually complex content. Outputs that use this feature incur pro-tier pricing. For information about feature limitations, see the AWS Elemental MediaConvert User Guide.

createJobTemplate_category :: Lens' CreateJobTemplate (Maybe Text)

Optional. A category for the job template you are creating

createJobTemplate_statusUpdateInterval :: Lens' CreateJobTemplate (Maybe StatusUpdateInterval)

Specify how often MediaConvert sends STATUS_UPDATE events to Amazon CloudWatch Events. Set the interval, in seconds, between status updates. MediaConvert sends an update at this interval from the time the service begins processing your job to the time it completes the transcode or encounters an error.

createJobTemplate_priority :: Lens' CreateJobTemplate (Maybe Int)

Specify the relative priority for this job. In any given queue, the service begins processing the job with the highest value first. When more than one job has the same priority, the service begins processing the job that you submitted first. If you don't specify a priority, the service uses the default value 0.

createJobTemplate_queue :: Lens' CreateJobTemplate (Maybe Text)

Optional. The queue that jobs created from this template are assigned to. If you don't specify this, jobs will go to the default queue.

createJobTemplate_tags :: Lens' CreateJobTemplate (Maybe (HashMap Text Text))

The tags that you want to add to the resource. You can tag resources with a key-value pair or with only a key.

createJobTemplate_description :: Lens' CreateJobTemplate (Maybe Text)

Optional. A description of the job template you are creating.

createJobTemplate_hopDestinations :: Lens' CreateJobTemplate (Maybe [HopDestination])

Optional. Use queue hopping to avoid overly long waits in the backlog of the queue that you submit your job to. Specify an alternate queue and the maximum time that your job will wait in the initial queue before hopping. For more information about this feature, see the AWS Elemental MediaConvert User Guide.

createJobTemplate_settings :: Lens' CreateJobTemplate JobTemplateSettings

JobTemplateSettings contains all the transcode settings saved in the template that will be applied to jobs created from it.

createJobTemplate_name :: Lens' CreateJobTemplate Text

The name of the job template you are creating.

Destructuring the Response

data CreateJobTemplateResponse

See: newCreateJobTemplateResponse smart constructor.

Constructors

CreateJobTemplateResponse'

Fields

jobTemplate :: Maybe JobTemplate

A job template is a pre-made set of encoding instructions that you can use to quickly create a job.

httpStatus :: Int

The response's http status code.

newCreateJobTemplateResponse

Arguments

:: Int
$sel:httpStatus:CreateJobTemplateResponse'

-> CreateJobTemplateResponse

Create a value of CreateJobTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobTemplate:CreateJobTemplateResponse', createJobTemplateResponse_jobTemplate - A job template is a pre-made set of encoding instructions that you can use to quickly create a job.

$sel:httpStatus:CreateJobTemplateResponse', createJobTemplateResponse_httpStatus - The response's http status code.

Response Lenses

createJobTemplateResponse_jobTemplate :: Lens' CreateJobTemplateResponse (Maybe JobTemplate)

A job template is a pre-made set of encoding instructions that you can use to quickly create a job.

createJobTemplateResponse_httpStatus :: Lens' CreateJobTemplateResponse Int

The response's http status code.