Network.AWS.MediaConvert.DescribeEndpoints

Description

Send an request with an empty body to the regional API endpoint to get your account API endpoint.

This operation returns paginated results.

Creating a Request

data DescribeEndpoints

DescribeEndpointsRequest

See: newDescribeEndpoints smart constructor.

Constructors

DescribeEndpoints'

Fields

mode :: Maybe DescribeEndpointsMode

Optional field, defaults to DEFAULT. Specify DEFAULT for this operation to return your endpoints if any exist, or to create an endpoint for you and return it if one doesn't already exist. Specify GET_ONLY to return your endpoints if any exist, or an empty list if none exist.

nextToken :: Maybe Text

Use this string, provided with the response to a previous request, to request the next batch of endpoints.

maxResults :: Maybe Int

Optional. Max number of endpoints, up to twenty, that will be returned at one time.

newDescribeEndpoints :: DescribeEndpoints

Create a value of DescribeEndpoints with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mode:DescribeEndpoints', describeEndpoints_mode - Optional field, defaults to DEFAULT. Specify DEFAULT for this operation to return your endpoints if any exist, or to create an endpoint for you and return it if one doesn't already exist. Specify GET_ONLY to return your endpoints if any exist, or an empty list if none exist.

$sel:nextToken:DescribeEndpoints', describeEndpoints_nextToken - Use this string, provided with the response to a previous request, to request the next batch of endpoints.

$sel:maxResults:DescribeEndpoints', describeEndpoints_maxResults - Optional. Max number of endpoints, up to twenty, that will be returned at one time.

Request Lenses

describeEndpoints_mode :: Lens' DescribeEndpoints (Maybe DescribeEndpointsMode)

Optional field, defaults to DEFAULT. Specify DEFAULT for this operation to return your endpoints if any exist, or to create an endpoint for you and return it if one doesn't already exist. Specify GET_ONLY to return your endpoints if any exist, or an empty list if none exist.

describeEndpoints_nextToken :: Lens' DescribeEndpoints (Maybe Text)

Use this string, provided with the response to a previous request, to request the next batch of endpoints.

describeEndpoints_maxResults :: Lens' DescribeEndpoints (Maybe Int)

Optional. Max number of endpoints, up to twenty, that will be returned at one time.

Destructuring the Response

data DescribeEndpointsResponse

See: newDescribeEndpointsResponse smart constructor.

Constructors

DescribeEndpointsResponse'

Fields

nextToken :: Maybe Text

Use this string to request the next batch of endpoints.

endpoints :: Maybe [Endpoint]

List of endpoints

httpStatus :: Int

The response's http status code.

newDescribeEndpointsResponse

Arguments

:: Int
$sel:httpStatus:DescribeEndpointsResponse'

-> DescribeEndpointsResponse

Create a value of DescribeEndpointsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeEndpoints', describeEndpointsResponse_nextToken - Use this string to request the next batch of endpoints.

$sel:endpoints:DescribeEndpointsResponse', describeEndpointsResponse_endpoints - List of endpoints

$sel:httpStatus:DescribeEndpointsResponse', describeEndpointsResponse_httpStatus - The response's http status code.

Response Lenses

describeEndpointsResponse_nextToken :: Lens' DescribeEndpointsResponse (Maybe Text)

Use this string to request the next batch of endpoints.

describeEndpointsResponse_endpoints :: Lens' DescribeEndpointsResponse (Maybe [Endpoint])

List of endpoints

describeEndpointsResponse_httpStatus :: Lens' DescribeEndpointsResponse Int

The response's http status code.