Network.AWS.MediaConvert.GetPolicy

Description

Retrieve the JSON for your policy.

Creating a Request

data GetPolicy

See: newGetPolicy smart constructor.

Constructors

GetPolicy'

newGetPolicy :: GetPolicy

Create a value of GetPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data GetPolicyResponse

See: newGetPolicyResponse smart constructor.

Constructors

GetPolicyResponse'

Fields

policy :: Maybe Policy

A policy configures behavior that you allow or disallow for your account. For information about MediaConvert policies, see the user guide at http://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html

httpStatus :: Int

The response's http status code.

newGetPolicyResponse

Arguments

:: Int
$sel:httpStatus:GetPolicyResponse'

-> GetPolicyResponse

Create a value of GetPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policy:GetPolicyResponse', getPolicyResponse_policy - A policy configures behavior that you allow or disallow for your account. For information about MediaConvert policies, see the user guide at http://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html

$sel:httpStatus:GetPolicyResponse', getPolicyResponse_httpStatus - The response's http status code.

Response Lenses

getPolicyResponse_policy :: Lens' GetPolicyResponse (Maybe Policy)

A policy configures behavior that you allow or disallow for your account. For information about MediaConvert policies, see the user guide at http://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html

getPolicyResponse_httpStatus :: Lens' GetPolicyResponse Int

The response's http status code.