Network.AWS.MediaConvert.ListJobTemplates

Description

Retrieve a JSON array of up to twenty of your job templates. This will return the templates themselves, not just a list of them. To retrieve the next twenty templates, use the nextToken string returned with the array

This operation returns paginated results.

Creating a Request

data ListJobTemplates

See: newListJobTemplates smart constructor.

Constructors

ListJobTemplates'

Fields

nextToken :: Maybe Text

Use this string, provided with the response to a previous request, to request the next batch of job templates.

listBy :: Maybe JobTemplateListBy

Optional. When you request a list of job templates, you can choose to list them alphabetically by NAME or chronologically by CREATION_DATE. If you don't specify, the service will list them by name.

maxResults :: Maybe Natural

Optional. Number of job templates, up to twenty, that will be returned at one time.

category :: Maybe Text

Optionally, specify a job template category to limit responses to only job templates from that category.

order :: Maybe Order

Optional. When you request lists of resources, you can specify whether they are sorted in ASCENDING or DESCENDING order. Default varies by resource.

newListJobTemplates :: ListJobTemplates

Create a value of ListJobTemplates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListJobTemplates', listJobTemplates_nextToken - Use this string, provided with the response to a previous request, to request the next batch of job templates.

$sel:listBy:ListJobTemplates', listJobTemplates_listBy - Optional. When you request a list of job templates, you can choose to list them alphabetically by NAME or chronologically by CREATION_DATE. If you don't specify, the service will list them by name.

$sel:maxResults:ListJobTemplates', listJobTemplates_maxResults - Optional. Number of job templates, up to twenty, that will be returned at one time.

$sel:category:ListJobTemplates', listJobTemplates_category - Optionally, specify a job template category to limit responses to only job templates from that category.

$sel:order:ListJobTemplates', listJobTemplates_order - Optional. When you request lists of resources, you can specify whether they are sorted in ASCENDING or DESCENDING order. Default varies by resource.

Request Lenses

listJobTemplates_nextToken :: Lens' ListJobTemplates (Maybe Text)

Use this string, provided with the response to a previous request, to request the next batch of job templates.

listJobTemplates_listBy :: Lens' ListJobTemplates (Maybe JobTemplateListBy)

Optional. When you request a list of job templates, you can choose to list them alphabetically by NAME or chronologically by CREATION_DATE. If you don't specify, the service will list them by name.

listJobTemplates_maxResults :: Lens' ListJobTemplates (Maybe Natural)

Optional. Number of job templates, up to twenty, that will be returned at one time.

listJobTemplates_category :: Lens' ListJobTemplates (Maybe Text)

Optionally, specify a job template category to limit responses to only job templates from that category.

listJobTemplates_order :: Lens' ListJobTemplates (Maybe Order)

Optional. When you request lists of resources, you can specify whether they are sorted in ASCENDING or DESCENDING order. Default varies by resource.

Destructuring the Response

data ListJobTemplatesResponse

See: newListJobTemplatesResponse smart constructor.

Constructors

ListJobTemplatesResponse'

Fields

nextToken :: Maybe Text

Use this string to request the next batch of job templates.

jobTemplates :: Maybe [JobTemplate]

List of Job templates.

httpStatus :: Int

The response's http status code.

newListJobTemplatesResponse

Arguments

:: Int
$sel:httpStatus:ListJobTemplatesResponse'

-> ListJobTemplatesResponse

Create a value of ListJobTemplatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListJobTemplates', listJobTemplatesResponse_nextToken - Use this string to request the next batch of job templates.

$sel:jobTemplates:ListJobTemplatesResponse', listJobTemplatesResponse_jobTemplates - List of Job templates.

$sel:httpStatus:ListJobTemplatesResponse', listJobTemplatesResponse_httpStatus - The response's http status code.

Response Lenses

listJobTemplatesResponse_nextToken :: Lens' ListJobTemplatesResponse (Maybe Text)

Use this string to request the next batch of job templates.

listJobTemplatesResponse_jobTemplates :: Lens' ListJobTemplatesResponse (Maybe [JobTemplate])

List of Job templates.

listJobTemplatesResponse_httpStatus :: Lens' ListJobTemplatesResponse Int

The response's http status code.