Network.AWS.MediaConvert.ListJobs

Description

Retrieve a JSON array of up to twenty of your most recently created jobs. This array includes in-process, completed, and errored jobs. This will return the jobs themselves, not just a list of the jobs. To retrieve the twenty next most recent jobs, use the nextToken string returned with the array.

This operation returns paginated results.

Creating a Request

data ListJobs

See: newListJobs smart constructor.

Constructors

ListJobs'

Fields

status :: Maybe JobStatus

Optional. A job's status can be SUBMITTED, PROGRESSING, COMPLETE, CANCELED, or ERROR.

nextToken :: Maybe Text

Optional. Use this string, provided with the response to a previous request, to request the next batch of jobs.

maxResults :: Maybe Natural

Optional. Number of jobs, up to twenty, that will be returned at one time.

queue :: Maybe Text

Optional. Provide a queue name to get back only jobs from that queue.

order :: Maybe Order

Optional. When you request lists of resources, you can specify whether they are sorted in ASCENDING or DESCENDING order. Default varies by resource.

newListJobs :: ListJobs

Create a value of ListJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ListJobs', listJobs_status - Optional. A job's status can be SUBMITTED, PROGRESSING, COMPLETE, CANCELED, or ERROR.

$sel:nextToken:ListJobs', listJobs_nextToken - Optional. Use this string, provided with the response to a previous request, to request the next batch of jobs.

$sel:maxResults:ListJobs', listJobs_maxResults - Optional. Number of jobs, up to twenty, that will be returned at one time.

$sel:queue:ListJobs', listJobs_queue - Optional. Provide a queue name to get back only jobs from that queue.

$sel:order:ListJobs', listJobs_order - Optional. When you request lists of resources, you can specify whether they are sorted in ASCENDING or DESCENDING order. Default varies by resource.

Request Lenses

listJobs_status :: Lens' ListJobs (Maybe JobStatus)

Optional. A job's status can be SUBMITTED, PROGRESSING, COMPLETE, CANCELED, or ERROR.

listJobs_nextToken :: Lens' ListJobs (Maybe Text)

Optional. Use this string, provided with the response to a previous request, to request the next batch of jobs.

listJobs_maxResults :: Lens' ListJobs (Maybe Natural)

Optional. Number of jobs, up to twenty, that will be returned at one time.

listJobs_queue :: Lens' ListJobs (Maybe Text)

Optional. Provide a queue name to get back only jobs from that queue.

listJobs_order :: Lens' ListJobs (Maybe Order)

Optional. When you request lists of resources, you can specify whether they are sorted in ASCENDING or DESCENDING order. Default varies by resource.

Destructuring the Response

data ListJobsResponse

See: newListJobsResponse smart constructor.

Constructors

ListJobsResponse'

Fields

nextToken :: Maybe Text

Use this string to request the next batch of jobs.

jobs :: Maybe [Job]

List of jobs

httpStatus :: Int

The response's http status code.

newListJobsResponse

Arguments

:: Int
$sel:httpStatus:ListJobsResponse'

-> ListJobsResponse

Create a value of ListJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListJobs', listJobsResponse_nextToken - Use this string to request the next batch of jobs.

$sel:jobs:ListJobsResponse', listJobsResponse_jobs - List of jobs

$sel:httpStatus:ListJobsResponse', listJobsResponse_httpStatus - The response's http status code.

Response Lenses

listJobsResponse_nextToken :: Lens' ListJobsResponse (Maybe Text)

Use this string to request the next batch of jobs.

listJobsResponse_jobs :: Lens' ListJobsResponse (Maybe [Job])

List of jobs

listJobsResponse_httpStatus :: Lens' ListJobsResponse Int

The response's http status code.