Network.AWS.MediaConvert.Types.Ac3Settings

Description

Documentation

data Ac3Settings

Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value AC3.

See: newAc3Settings smart constructor.

Constructors

Ac3Settings'

Fields

dialnorm :: Maybe Natural

Sets the dialnorm for the output. If blank and input audio is Dolby Digital, dialnorm will be passed through.

codingMode :: Maybe Ac3CodingMode

Dolby Digital coding mode. Determines number of channels.

dynamicRangeCompressionProfile :: Maybe Ac3DynamicRangeCompressionProfile

When you want to add Dolby dynamic range compression (DRC) signaling to your output stream, we recommend that you use the mode-specific settings instead of Dynamic range compression profile (DynamicRangeCompressionProfile). The mode-specific settings are Dynamic range compression profile, line mode (dynamicRangeCompressionLine) and Dynamic range compression profile, RF mode (dynamicRangeCompressionRf). Note that when you specify values for all three settings, MediaConvert ignores the value of this setting in favor of the mode-specific settings. If you do use this setting instead of the mode-specific settings, choose None (NONE) to leave out DRC signaling. Keep the default Film standard (FILM_STANDARD) to set the profile to Dolby's film standard profile for all operating modes.

lfeFilter :: Maybe Ac3LfeFilter

Applies a 120Hz lowpass filter to the LFE channel prior to encoding. Only valid with 3_2_LFE coding mode.

sampleRate :: Maybe Natural

This value is always 48000. It represents the sample rate in Hz.

dynamicRangeCompressionRf :: Maybe Ac3DynamicRangeCompressionRf

Choose the Dolby Digital dynamic range control (DRC) profile that MediaConvert uses when encoding the metadata in the Dolby Digital stream for the RF operating mode. Related setting: When you use this setting, MediaConvert ignores any value you provide for Dynamic range compression profile (DynamicRangeCompressionProfile). For information about the Dolby Digital DRC operating modes and profiles, see the Dynamic Range Control chapter of the Dolby Metadata Guide at https://developer.dolby.com/globalassets/professional/documents/dolby-metadata-guide.pdf.

bitstreamMode :: Maybe Ac3BitstreamMode

Specify the bitstream mode for the AC-3 stream that the encoder emits. For more information about the AC3 bitstream mode, see ATSC A/52-2012 (Annex E).

bitrate :: Maybe Natural

Specify the average bitrate in bits per second. Valid bitrates depend on the coding mode.

dynamicRangeCompressionLine :: Maybe Ac3DynamicRangeCompressionLine

Choose the Dolby Digital dynamic range control (DRC) profile that MediaConvert uses when encoding the metadata in the Dolby Digital stream for the line operating mode. Related setting: When you use this setting, MediaConvert ignores any value you provide for Dynamic range compression profile (DynamicRangeCompressionProfile). For information about the Dolby Digital DRC operating modes and profiles, see the Dynamic Range Control chapter of the Dolby Metadata Guide at https://developer.dolby.com/globalassets/professional/documents/dolby-metadata-guide.pdf.

metadataControl :: Maybe Ac3MetadataControl

When set to FOLLOW_INPUT, encoder metadata will be sourced from the DD, DD+, or DolbyE decoder that supplied this audio data. If audio was not supplied from one of these streams, then the static metadata settings will be used.

newAc3Settings :: Ac3Settings

Create a value of Ac3Settings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dialnorm:Ac3Settings', ac3Settings_dialnorm - Sets the dialnorm for the output. If blank and input audio is Dolby Digital, dialnorm will be passed through.

$sel:codingMode:Ac3Settings', ac3Settings_codingMode - Dolby Digital coding mode. Determines number of channels.

$sel:dynamicRangeCompressionProfile:Ac3Settings', ac3Settings_dynamicRangeCompressionProfile - When you want to add Dolby dynamic range compression (DRC) signaling to your output stream, we recommend that you use the mode-specific settings instead of Dynamic range compression profile (DynamicRangeCompressionProfile). The mode-specific settings are Dynamic range compression profile, line mode (dynamicRangeCompressionLine) and Dynamic range compression profile, RF mode (dynamicRangeCompressionRf). Note that when you specify values for all three settings, MediaConvert ignores the value of this setting in favor of the mode-specific settings. If you do use this setting instead of the mode-specific settings, choose None (NONE) to leave out DRC signaling. Keep the default Film standard (FILM_STANDARD) to set the profile to Dolby's film standard profile for all operating modes.

$sel:lfeFilter:Ac3Settings', ac3Settings_lfeFilter - Applies a 120Hz lowpass filter to the LFE channel prior to encoding. Only valid with 3_2_LFE coding mode.

$sel:sampleRate:Ac3Settings', ac3Settings_sampleRate - This value is always 48000. It represents the sample rate in Hz.

$sel:dynamicRangeCompressionRf:Ac3Settings', ac3Settings_dynamicRangeCompressionRf - Choose the Dolby Digital dynamic range control (DRC) profile that MediaConvert uses when encoding the metadata in the Dolby Digital stream for the RF operating mode. Related setting: When you use this setting, MediaConvert ignores any value you provide for Dynamic range compression profile (DynamicRangeCompressionProfile). For information about the Dolby Digital DRC operating modes and profiles, see the Dynamic Range Control chapter of the Dolby Metadata Guide at https://developer.dolby.com/globalassets/professional/documents/dolby-metadata-guide.pdf.

$sel:bitstreamMode:Ac3Settings', ac3Settings_bitstreamMode - Specify the bitstream mode for the AC-3 stream that the encoder emits. For more information about the AC3 bitstream mode, see ATSC A/52-2012 (Annex E).

$sel:bitrate:Ac3Settings', ac3Settings_bitrate - Specify the average bitrate in bits per second. Valid bitrates depend on the coding mode.

$sel:dynamicRangeCompressionLine:Ac3Settings', ac3Settings_dynamicRangeCompressionLine - Choose the Dolby Digital dynamic range control (DRC) profile that MediaConvert uses when encoding the metadata in the Dolby Digital stream for the line operating mode. Related setting: When you use this setting, MediaConvert ignores any value you provide for Dynamic range compression profile (DynamicRangeCompressionProfile). For information about the Dolby Digital DRC operating modes and profiles, see the Dynamic Range Control chapter of the Dolby Metadata Guide at https://developer.dolby.com/globalassets/professional/documents/dolby-metadata-guide.pdf.

$sel:metadataControl:Ac3Settings', ac3Settings_metadataControl - When set to FOLLOW_INPUT, encoder metadata will be sourced from the DD, DD+, or DolbyE decoder that supplied this audio data. If audio was not supplied from one of these streams, then the static metadata settings will be used.

ac3Settings_dialnorm :: Lens' Ac3Settings (Maybe Natural)

Sets the dialnorm for the output. If blank and input audio is Dolby Digital, dialnorm will be passed through.

ac3Settings_codingMode :: Lens' Ac3Settings (Maybe Ac3CodingMode)

Dolby Digital coding mode. Determines number of channels.

ac3Settings_dynamicRangeCompressionProfile :: Lens' Ac3Settings (Maybe Ac3DynamicRangeCompressionProfile)

When you want to add Dolby dynamic range compression (DRC) signaling to your output stream, we recommend that you use the mode-specific settings instead of Dynamic range compression profile (DynamicRangeCompressionProfile). The mode-specific settings are Dynamic range compression profile, line mode (dynamicRangeCompressionLine) and Dynamic range compression profile, RF mode (dynamicRangeCompressionRf). Note that when you specify values for all three settings, MediaConvert ignores the value of this setting in favor of the mode-specific settings. If you do use this setting instead of the mode-specific settings, choose None (NONE) to leave out DRC signaling. Keep the default Film standard (FILM_STANDARD) to set the profile to Dolby's film standard profile for all operating modes.

ac3Settings_lfeFilter :: Lens' Ac3Settings (Maybe Ac3LfeFilter)

Applies a 120Hz lowpass filter to the LFE channel prior to encoding. Only valid with 3_2_LFE coding mode.

ac3Settings_sampleRate :: Lens' Ac3Settings (Maybe Natural)

This value is always 48000. It represents the sample rate in Hz.

ac3Settings_dynamicRangeCompressionRf :: Lens' Ac3Settings (Maybe Ac3DynamicRangeCompressionRf)

Choose the Dolby Digital dynamic range control (DRC) profile that MediaConvert uses when encoding the metadata in the Dolby Digital stream for the RF operating mode. Related setting: When you use this setting, MediaConvert ignores any value you provide for Dynamic range compression profile (DynamicRangeCompressionProfile). For information about the Dolby Digital DRC operating modes and profiles, see the Dynamic Range Control chapter of the Dolby Metadata Guide at https://developer.dolby.com/globalassets/professional/documents/dolby-metadata-guide.pdf.

ac3Settings_bitstreamMode :: Lens' Ac3Settings (Maybe Ac3BitstreamMode)

Specify the bitstream mode for the AC-3 stream that the encoder emits. For more information about the AC3 bitstream mode, see ATSC A/52-2012 (Annex E).

ac3Settings_bitrate :: Lens' Ac3Settings (Maybe Natural)

Specify the average bitrate in bits per second. Valid bitrates depend on the coding mode.

ac3Settings_dynamicRangeCompressionLine :: Lens' Ac3Settings (Maybe Ac3DynamicRangeCompressionLine)

Choose the Dolby Digital dynamic range control (DRC) profile that MediaConvert uses when encoding the metadata in the Dolby Digital stream for the line operating mode. Related setting: When you use this setting, MediaConvert ignores any value you provide for Dynamic range compression profile (DynamicRangeCompressionProfile). For information about the Dolby Digital DRC operating modes and profiles, see the Dynamic Range Control chapter of the Dolby Metadata Guide at https://developer.dolby.com/globalassets/professional/documents/dolby-metadata-guide.pdf.

ac3Settings_metadataControl :: Lens' Ac3Settings (Maybe Ac3MetadataControl)

When set to FOLLOW_INPUT, encoder metadata will be sourced from the DD, DD+, or DolbyE decoder that supplied this audio data. If audio was not supplied from one of these streams, then the static metadata settings will be used.