Network.AWS.MediaConvert.Types.AudioCodecSettings

Description

Documentation

data AudioCodecSettings

Settings related to audio encoding. The settings in this group vary depending on the value that you choose for your audio codec.

See: newAudioCodecSettings smart constructor.

Constructors

AudioCodecSettings'

Fields

ac3Settings :: Maybe Ac3Settings

Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value AC3.

vorbisSettings :: Maybe VorbisSettings

Required when you set Codec, under AudioDescriptions>CodecSettings, to the value Vorbis.

codec :: Maybe AudioCodec

Choose the audio codec for this output. Note that the option Dolby Digital passthrough (PASSTHROUGH) applies only to Dolby Digital and Dolby Digital Plus audio inputs. Make sure that you choose a codec that's supported with your output container: https://docs.aws.amazon.com/mediaconvert/latest/ug/reference-codecs-containers.html#reference-codecs-containers-output-audio For audio-only outputs, make sure that both your input audio codec and your output audio codec are supported for audio-only workflows. For more information, see: https://docs.aws.amazon.com/mediaconvert/latest/ug/reference-codecs-containers-input.html#reference-codecs-containers-input-audio-only and https://docs.aws.amazon.com/mediaconvert/latest/ug/reference-codecs-containers.html#audio-only-output

mp3Settings :: Maybe Mp3Settings

Required when you set Codec, under AudioDescriptions>CodecSettings, to the value MP3.

mp2Settings :: Maybe Mp2Settings

Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value MP2.

opusSettings :: Maybe OpusSettings

Required when you set Codec, under AudioDescriptions>CodecSettings, to the value OPUS.

eac3Settings :: Maybe Eac3Settings

Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value EAC3.

aacSettings :: Maybe AacSettings

Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value AAC. The service accepts one of two mutually exclusive groups of AAC settings--VBR and CBR. To select one of these modes, set the value of Bitrate control mode (rateControlMode) to "VBR" or "CBR". In VBR mode, you control the audio quality with the setting VBR quality (vbrQuality). In CBR mode, you use the setting Bitrate (bitrate). Defaults and valid values depend on the rate control mode.

aiffSettings :: Maybe AiffSettings

Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value AIFF.

eac3AtmosSettings :: Maybe Eac3AtmosSettings

Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value EAC3_ATMOS.

wavSettings :: Maybe WavSettings

Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value WAV.

newAudioCodecSettings :: AudioCodecSettings

Create a value of AudioCodecSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ac3Settings:AudioCodecSettings', audioCodecSettings_ac3Settings - Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value AC3.

$sel:vorbisSettings:AudioCodecSettings', audioCodecSettings_vorbisSettings - Required when you set Codec, under AudioDescriptions>CodecSettings, to the value Vorbis.

$sel:codec:AudioCodecSettings', audioCodecSettings_codec - Choose the audio codec for this output. Note that the option Dolby Digital passthrough (PASSTHROUGH) applies only to Dolby Digital and Dolby Digital Plus audio inputs. Make sure that you choose a codec that's supported with your output container: https://docs.aws.amazon.com/mediaconvert/latest/ug/reference-codecs-containers.html#reference-codecs-containers-output-audio For audio-only outputs, make sure that both your input audio codec and your output audio codec are supported for audio-only workflows. For more information, see: https://docs.aws.amazon.com/mediaconvert/latest/ug/reference-codecs-containers-input.html#reference-codecs-containers-input-audio-only and https://docs.aws.amazon.com/mediaconvert/latest/ug/reference-codecs-containers.html#audio-only-output

$sel:mp3Settings:AudioCodecSettings', audioCodecSettings_mp3Settings - Required when you set Codec, under AudioDescriptions>CodecSettings, to the value MP3.

$sel:mp2Settings:AudioCodecSettings', audioCodecSettings_mp2Settings - Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value MP2.

$sel:opusSettings:AudioCodecSettings', audioCodecSettings_opusSettings - Required when you set Codec, under AudioDescriptions>CodecSettings, to the value OPUS.

$sel:eac3Settings:AudioCodecSettings', audioCodecSettings_eac3Settings - Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value EAC3.

$sel:aacSettings:AudioCodecSettings', audioCodecSettings_aacSettings - Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value AAC. The service accepts one of two mutually exclusive groups of AAC settings--VBR and CBR. To select one of these modes, set the value of Bitrate control mode (rateControlMode) to "VBR" or "CBR". In VBR mode, you control the audio quality with the setting VBR quality (vbrQuality). In CBR mode, you use the setting Bitrate (bitrate). Defaults and valid values depend on the rate control mode.

$sel:aiffSettings:AudioCodecSettings', audioCodecSettings_aiffSettings - Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value AIFF.

$sel:eac3AtmosSettings:AudioCodecSettings', audioCodecSettings_eac3AtmosSettings - Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value EAC3_ATMOS.

$sel:wavSettings:AudioCodecSettings', audioCodecSettings_wavSettings - Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value WAV.

audioCodecSettings_ac3Settings :: Lens' AudioCodecSettings (Maybe Ac3Settings)

Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value AC3.

audioCodecSettings_vorbisSettings :: Lens' AudioCodecSettings (Maybe VorbisSettings)

Required when you set Codec, under AudioDescriptions>CodecSettings, to the value Vorbis.

audioCodecSettings_codec :: Lens' AudioCodecSettings (Maybe AudioCodec)

Choose the audio codec for this output. Note that the option Dolby Digital passthrough (PASSTHROUGH) applies only to Dolby Digital and Dolby Digital Plus audio inputs. Make sure that you choose a codec that's supported with your output container: https://docs.aws.amazon.com/mediaconvert/latest/ug/reference-codecs-containers.html#reference-codecs-containers-output-audio For audio-only outputs, make sure that both your input audio codec and your output audio codec are supported for audio-only workflows. For more information, see: https://docs.aws.amazon.com/mediaconvert/latest/ug/reference-codecs-containers-input.html#reference-codecs-containers-input-audio-only and https://docs.aws.amazon.com/mediaconvert/latest/ug/reference-codecs-containers.html#audio-only-output

audioCodecSettings_mp3Settings :: Lens' AudioCodecSettings (Maybe Mp3Settings)

Required when you set Codec, under AudioDescriptions>CodecSettings, to the value MP3.

audioCodecSettings_mp2Settings :: Lens' AudioCodecSettings (Maybe Mp2Settings)

Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value MP2.

audioCodecSettings_opusSettings :: Lens' AudioCodecSettings (Maybe OpusSettings)

Required when you set Codec, under AudioDescriptions>CodecSettings, to the value OPUS.

audioCodecSettings_eac3Settings :: Lens' AudioCodecSettings (Maybe Eac3Settings)

Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value EAC3.

audioCodecSettings_aacSettings :: Lens' AudioCodecSettings (Maybe AacSettings)

Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value AAC. The service accepts one of two mutually exclusive groups of AAC settings--VBR and CBR. To select one of these modes, set the value of Bitrate control mode (rateControlMode) to "VBR" or "CBR". In VBR mode, you control the audio quality with the setting VBR quality (vbrQuality). In CBR mode, you use the setting Bitrate (bitrate). Defaults and valid values depend on the rate control mode.

audioCodecSettings_aiffSettings :: Lens' AudioCodecSettings (Maybe AiffSettings)

Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value AIFF.

audioCodecSettings_eac3AtmosSettings :: Lens' AudioCodecSettings (Maybe Eac3AtmosSettings)

Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value EAC3_ATMOS.

audioCodecSettings_wavSettings :: Lens' AudioCodecSettings (Maybe WavSettings)

Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value WAV.