Network.AWS.MediaConvert.Types.AudioSelector

Description

Documentation

data AudioSelector

Use Audio selectors (AudioSelectors) to specify a track or set of tracks from the input that you will use in your outputs. You can use multiple Audio selectors per input.

See: newAudioSelector smart constructor.

Constructors

AudioSelector'

Fields

languageCode :: Maybe LanguageCode

Selects a specific language code from within an audio source.

tracks :: Maybe [Natural]

Identify a track from the input audio to include in this selector by entering the track index number. To include several tracks in a single audio selector, specify multiple tracks as follows. Using the console, enter a comma-separated list. For examle, type "1,2,3" to include tracks 1 through 3. Specifying directly in your JSON job file, provide the track numbers in an array. For example, "tracks": [1,2,3].

programSelection :: Maybe Natural

Use this setting for input streams that contain Dolby E, to have the service extract specific program data from the track. To select multiple programs, create multiple selectors with the same Track and different Program numbers. In the console, this setting is visible when you set Selector type to Track. Choose the program number from the dropdown list. If you are sending a JSON file, provide the program ID, which is part of the audio metadata. If your input file has incorrect metadata, you can choose All channels instead of a program number to have the service ignore the program IDs and include all the programs in the track.

customLanguageCode :: Maybe Text

Selects a specific language code from within an audio source, using the ISO 639-2 or ISO 639-3 three-letter language code

defaultSelection :: Maybe AudioDefaultSelection

Enable this setting on one audio selector to set it as the default for the job. The service uses this default for outputs where it can't find the specified input audio. If you don't set a default, those outputs have no audio.

selectorType :: Maybe AudioSelectorType

Specifies the type of the audio selector.

remixSettings :: Maybe RemixSettings

Use these settings to reorder the audio channels of one input to match those of another input. This allows you to combine the two files into a single output, one after the other.

pids :: Maybe [Natural]

Selects a specific PID from within an audio source (e.g. 257 selects PID 0x101).

externalAudioFileInput :: Maybe Text

Specifies audio data from an external file source.

offset :: Maybe Int

Specifies a time delta in milliseconds to offset the audio from the input video.

hlsRenditionGroupSettings :: Maybe HlsRenditionGroupSettings

Settings specific to audio sources in an HLS alternate rendition group. Specify the properties (renditionGroupId, renditionName or renditionLanguageCode) to identify the unique audio track among the alternative rendition groups present in the HLS manifest. If no unique track is found, or multiple tracks match the properties provided, the job fails. If no properties in hlsRenditionGroupSettings are specified, the default audio track within the video segment is chosen. If there is no audio within video segment, the alternative audio with DEFAULT=YES is chosen instead.

newAudioSelector :: AudioSelector

Create a value of AudioSelector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:AudioSelector', audioSelector_languageCode - Selects a specific language code from within an audio source.

$sel:tracks:AudioSelector', audioSelector_tracks - Identify a track from the input audio to include in this selector by entering the track index number. To include several tracks in a single audio selector, specify multiple tracks as follows. Using the console, enter a comma-separated list. For examle, type "1,2,3" to include tracks 1 through 3. Specifying directly in your JSON job file, provide the track numbers in an array. For example, "tracks": [1,2,3].

$sel:programSelection:AudioSelector', audioSelector_programSelection - Use this setting for input streams that contain Dolby E, to have the service extract specific program data from the track. To select multiple programs, create multiple selectors with the same Track and different Program numbers. In the console, this setting is visible when you set Selector type to Track. Choose the program number from the dropdown list. If you are sending a JSON file, provide the program ID, which is part of the audio metadata. If your input file has incorrect metadata, you can choose All channels instead of a program number to have the service ignore the program IDs and include all the programs in the track.

$sel:customLanguageCode:AudioSelector', audioSelector_customLanguageCode - Selects a specific language code from within an audio source, using the ISO 639-2 or ISO 639-3 three-letter language code

$sel:defaultSelection:AudioSelector', audioSelector_defaultSelection - Enable this setting on one audio selector to set it as the default for the job. The service uses this default for outputs where it can't find the specified input audio. If you don't set a default, those outputs have no audio.

$sel:selectorType:AudioSelector', audioSelector_selectorType - Specifies the type of the audio selector.

$sel:remixSettings:AudioSelector', audioSelector_remixSettings - Use these settings to reorder the audio channels of one input to match those of another input. This allows you to combine the two files into a single output, one after the other.

$sel:pids:AudioSelector', audioSelector_pids - Selects a specific PID from within an audio source (e.g. 257 selects PID 0x101).

$sel:externalAudioFileInput:AudioSelector', audioSelector_externalAudioFileInput - Specifies audio data from an external file source.

$sel:offset:AudioSelector', audioSelector_offset - Specifies a time delta in milliseconds to offset the audio from the input video.

$sel:hlsRenditionGroupSettings:AudioSelector', audioSelector_hlsRenditionGroupSettings - Settings specific to audio sources in an HLS alternate rendition group. Specify the properties (renditionGroupId, renditionName or renditionLanguageCode) to identify the unique audio track among the alternative rendition groups present in the HLS manifest. If no unique track is found, or multiple tracks match the properties provided, the job fails. If no properties in hlsRenditionGroupSettings are specified, the default audio track within the video segment is chosen. If there is no audio within video segment, the alternative audio with DEFAULT=YES is chosen instead.

audioSelector_languageCode :: Lens' AudioSelector (Maybe LanguageCode)

Selects a specific language code from within an audio source.

audioSelector_tracks :: Lens' AudioSelector (Maybe [Natural])

Identify a track from the input audio to include in this selector by entering the track index number. To include several tracks in a single audio selector, specify multiple tracks as follows. Using the console, enter a comma-separated list. For examle, type "1,2,3" to include tracks 1 through 3. Specifying directly in your JSON job file, provide the track numbers in an array. For example, "tracks": [1,2,3].

audioSelector_programSelection :: Lens' AudioSelector (Maybe Natural)

Use this setting for input streams that contain Dolby E, to have the service extract specific program data from the track. To select multiple programs, create multiple selectors with the same Track and different Program numbers. In the console, this setting is visible when you set Selector type to Track. Choose the program number from the dropdown list. If you are sending a JSON file, provide the program ID, which is part of the audio metadata. If your input file has incorrect metadata, you can choose All channels instead of a program number to have the service ignore the program IDs and include all the programs in the track.

audioSelector_customLanguageCode :: Lens' AudioSelector (Maybe Text)

Selects a specific language code from within an audio source, using the ISO 639-2 or ISO 639-3 three-letter language code

audioSelector_defaultSelection :: Lens' AudioSelector (Maybe AudioDefaultSelection)

Enable this setting on one audio selector to set it as the default for the job. The service uses this default for outputs where it can't find the specified input audio. If you don't set a default, those outputs have no audio.

audioSelector_selectorType :: Lens' AudioSelector (Maybe AudioSelectorType)

Specifies the type of the audio selector.

audioSelector_remixSettings :: Lens' AudioSelector (Maybe RemixSettings)

Use these settings to reorder the audio channels of one input to match those of another input. This allows you to combine the two files into a single output, one after the other.

audioSelector_pids :: Lens' AudioSelector (Maybe [Natural])

Selects a specific PID from within an audio source (e.g. 257 selects PID 0x101).

audioSelector_externalAudioFileInput :: Lens' AudioSelector (Maybe Text)

Specifies audio data from an external file source.

audioSelector_offset :: Lens' AudioSelector (Maybe Int)

Specifies a time delta in milliseconds to offset the audio from the input video.

audioSelector_hlsRenditionGroupSettings :: Lens' AudioSelector (Maybe HlsRenditionGroupSettings)

Settings specific to audio sources in an HLS alternate rendition group. Specify the properties (renditionGroupId, renditionName or renditionLanguageCode) to identify the unique audio track among the alternative rendition groups present in the HLS manifest. If no unique track is found, or multiple tracks match the properties provided, the job fails. If no properties in hlsRenditionGroupSettings are specified, the default audio track within the video segment is chosen. If there is no audio within video segment, the alternative audio with DEFAULT=YES is chosen instead.