Network.AWS.MediaConvert.Types.CaptionDestinationType

Description

Documentation

newtype CaptionDestinationType

Specify the format for this set of captions on this output. The default format is embedded without SCTE-20. Note that your choice of video output container constrains your choice of output captions format. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/captions-support-tables.html. If you are using SCTE-20 and you want to create an output that complies with the SCTE-43 spec, choose SCTE-20 plus embedded (SCTE20_PLUS_EMBEDDED). To create a non-compliant output where the embedded captions come first, choose Embedded plus SCTE-20 (EMBEDDED_PLUS_SCTE20).

Constructors

CaptionDestinationType'

Fields

fromCaptionDestinationType :: Text

Bundled Patterns

pattern CaptionDestinationType_BURN_IN :: CaptionDestinationType

pattern CaptionDestinationType_DVB_SUB :: CaptionDestinationType

pattern CaptionDestinationType_EMBEDDED :: CaptionDestinationType

pattern CaptionDestinationType_EMBEDDED_PLUS_SCTE20 :: CaptionDestinationType

pattern CaptionDestinationType_IMSC :: CaptionDestinationType

pattern CaptionDestinationType_SCC :: CaptionDestinationType

pattern CaptionDestinationType_SCTE20_PLUS_EMBEDDED :: CaptionDestinationType

pattern CaptionDestinationType_SMI :: CaptionDestinationType

pattern CaptionDestinationType_SRT :: CaptionDestinationType

pattern CaptionDestinationType_TELETEXT :: CaptionDestinationType

pattern CaptionDestinationType_TTML :: CaptionDestinationType

pattern CaptionDestinationType_WEBVTT :: CaptionDestinationType