Network.AWS.MediaConvert.Types.CaptionSourceFramerate

Description

Documentation

data CaptionSourceFramerate

Ignore this setting unless your input captions format is SCC. To have the service compensate for differing frame rates between your input captions and input video, specify the frame rate of the captions file. Specify this value as a fraction, using the settings Framerate numerator (framerateNumerator) and Framerate denominator (framerateDenominator). For example, you might specify 24 / 1 for 24 fps, 25 / 1 for 25 fps, 24000 / 1001 for 23.976 fps, or 30000 / 1001 for 29.97 fps.

See: newCaptionSourceFramerate smart constructor.

Constructors

CaptionSourceFramerate'

Fields

framerateNumerator :: Maybe Natural

Specify the numerator of the fraction that represents the frame rate for the setting Caption source frame rate (CaptionSourceFramerate). Use this setting along with the setting Framerate denominator (framerateDenominator).

framerateDenominator :: Maybe Natural

Specify the denominator of the fraction that represents the frame rate for the setting Caption source frame rate (CaptionSourceFramerate). Use this setting along with the setting Framerate numerator (framerateNumerator).

newCaptionSourceFramerate :: CaptionSourceFramerate

Create a value of CaptionSourceFramerate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:framerateNumerator:CaptionSourceFramerate', captionSourceFramerate_framerateNumerator - Specify the numerator of the fraction that represents the frame rate for the setting Caption source frame rate (CaptionSourceFramerate). Use this setting along with the setting Framerate denominator (framerateDenominator).

$sel:framerateDenominator:CaptionSourceFramerate', captionSourceFramerate_framerateDenominator - Specify the denominator of the fraction that represents the frame rate for the setting Caption source frame rate (CaptionSourceFramerate). Use this setting along with the setting Framerate numerator (framerateNumerator).

captionSourceFramerate_framerateNumerator :: Lens' CaptionSourceFramerate (Maybe Natural)

Specify the numerator of the fraction that represents the frame rate for the setting Caption source frame rate (CaptionSourceFramerate). Use this setting along with the setting Framerate denominator (framerateDenominator).

captionSourceFramerate_framerateDenominator :: Lens' CaptionSourceFramerate (Maybe Natural)

Specify the denominator of the fraction that represents the frame rate for the setting Caption source frame rate (CaptionSourceFramerate). Use this setting along with the setting Framerate numerator (framerateNumerator).