Network.AWS.MediaConvert.Types.ContainerSettings

Description

Documentation

data ContainerSettings

Container specific settings.

See: newContainerSettings smart constructor.

Constructors

ContainerSettings'

Fields

container :: Maybe ContainerType

Container for this output. Some containers require a container settings object. If not specified, the default object will be created.

mp4Settings :: Maybe Mp4Settings

These settings relate to your MP4 output container. You can create audio only outputs with this container. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/supported-codecs-containers-audio-only.html#output-codecs-and-containers-supported-for-audio-only.

mpdSettings :: Maybe MpdSettings

These settings relate to the fragmented MP4 container for the segments in your DASH outputs.

f4vSettings :: Maybe F4vSettings

Settings for F4v container

mxfSettings :: Maybe MxfSettings

These settings relate to your MXF output container.

movSettings :: Maybe MovSettings

These settings relate to your QuickTime MOV output container.

cmfcSettings :: Maybe CmfcSettings

These settings relate to the fragmented MP4 container for the segments in your CMAF outputs.

m3u8Settings :: Maybe M3u8Settings

These settings relate to the MPEG-2 transport stream (MPEG2-TS) container for the MPEG2-TS segments in your HLS outputs.

m2tsSettings :: Maybe M2tsSettings

MPEG-2 TS container settings. These apply to outputs in a File output group when the output's container (ContainerType) is MPEG-2 Transport Stream (M2TS). In these assets, data is organized by the program map table (PMT). Each transport stream program contains subsets of data, including audio, video, and metadata. Each of these subsets of data has a numerical label called a packet identifier (PID). Each transport stream program corresponds to one MediaConvert output. The PMT lists the types of data in a program along with their PID. Downstream systems and players use the program map table to look up the PID for each type of data it accesses and then uses the PIDs to locate specific data within the asset.

newContainerSettings :: ContainerSettings

Create a value of ContainerSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:container:ContainerSettings', containerSettings_container - Container for this output. Some containers require a container settings object. If not specified, the default object will be created.

$sel:mp4Settings:ContainerSettings', containerSettings_mp4Settings - These settings relate to your MP4 output container. You can create audio only outputs with this container. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/supported-codecs-containers-audio-only.html#output-codecs-and-containers-supported-for-audio-only.

$sel:mpdSettings:ContainerSettings', containerSettings_mpdSettings - These settings relate to the fragmented MP4 container for the segments in your DASH outputs.

$sel:f4vSettings:ContainerSettings', containerSettings_f4vSettings - Settings for F4v container

$sel:mxfSettings:ContainerSettings', containerSettings_mxfSettings - These settings relate to your MXF output container.

$sel:movSettings:ContainerSettings', containerSettings_movSettings - These settings relate to your QuickTime MOV output container.

$sel:cmfcSettings:ContainerSettings', containerSettings_cmfcSettings - These settings relate to the fragmented MP4 container for the segments in your CMAF outputs.

$sel:m3u8Settings:ContainerSettings', containerSettings_m3u8Settings - These settings relate to the MPEG-2 transport stream (MPEG2-TS) container for the MPEG2-TS segments in your HLS outputs.

$sel:m2tsSettings:ContainerSettings', containerSettings_m2tsSettings - MPEG-2 TS container settings. These apply to outputs in a File output group when the output's container (ContainerType) is MPEG-2 Transport Stream (M2TS). In these assets, data is organized by the program map table (PMT). Each transport stream program contains subsets of data, including audio, video, and metadata. Each of these subsets of data has a numerical label called a packet identifier (PID). Each transport stream program corresponds to one MediaConvert output. The PMT lists the types of data in a program along with their PID. Downstream systems and players use the program map table to look up the PID for each type of data it accesses and then uses the PIDs to locate specific data within the asset.

containerSettings_container :: Lens' ContainerSettings (Maybe ContainerType)

Container for this output. Some containers require a container settings object. If not specified, the default object will be created.

containerSettings_mp4Settings :: Lens' ContainerSettings (Maybe Mp4Settings)

These settings relate to your MP4 output container. You can create audio only outputs with this container. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/supported-codecs-containers-audio-only.html#output-codecs-and-containers-supported-for-audio-only.

containerSettings_mpdSettings :: Lens' ContainerSettings (Maybe MpdSettings)

These settings relate to the fragmented MP4 container for the segments in your DASH outputs.

containerSettings_f4vSettings :: Lens' ContainerSettings (Maybe F4vSettings)

Settings for F4v container

containerSettings_mxfSettings :: Lens' ContainerSettings (Maybe MxfSettings)

These settings relate to your MXF output container.

containerSettings_movSettings :: Lens' ContainerSettings (Maybe MovSettings)

These settings relate to your QuickTime MOV output container.

containerSettings_cmfcSettings :: Lens' ContainerSettings (Maybe CmfcSettings)

These settings relate to the fragmented MP4 container for the segments in your CMAF outputs.

containerSettings_m3u8Settings :: Lens' ContainerSettings (Maybe M3u8Settings)

These settings relate to the MPEG-2 transport stream (MPEG2-TS) container for the MPEG2-TS segments in your HLS outputs.

containerSettings_m2tsSettings :: Lens' ContainerSettings (Maybe M2tsSettings)

MPEG-2 TS container settings. These apply to outputs in a File output group when the output's container (ContainerType) is MPEG-2 Transport Stream (M2TS). In these assets, data is organized by the program map table (PMT). Each transport stream program contains subsets of data, including audio, video, and metadata. Each of these subsets of data has a numerical label called a packet identifier (PID). Each transport stream program corresponds to one MediaConvert output. The PMT lists the types of data in a program along with their PID. Downstream systems and players use the program map table to look up the PID for each type of data it accesses and then uses the PIDs to locate specific data within the asset.