Network.AWS.MediaConvert.Types.Eac3Settings

Description

Documentation

data Eac3Settings

Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value EAC3.

See: newEac3Settings smart constructor.

Constructors

Eac3Settings'

Fields

stereoDownmix :: Maybe Eac3StereoDownmix

Choose how the service does stereo downmixing. This setting only applies if you keep the default value of 3/2 - L, R, C, Ls, Rs (CODING_MODE_3_2) for the setting Coding mode (Eac3CodingMode). If you choose a different value for Coding mode, the service ignores Stereo downmix (Eac3StereoDownmix).

loRoCenterMixLevel :: Maybe Double

Specify a value for the following Dolby Digital Plus setting: Left only/Right only center mix (Lo/Ro center). MediaConvert uses this value for downmixing. How the service uses this value depends on the value that you choose for Stereo downmix (Eac3StereoDownmix). Valid values: 3.0, 1.5, 0.0, -1.5, -3.0, -4.5, -6.0, and -60. The value -60 mutes the channel. This setting applies only if you keep the default value of 3/2 - L, R, C, Ls, Rs (CODING_MODE_3_2) for the setting Coding mode (Eac3CodingMode). If you choose a different value for Coding mode, the service ignores Left only/Right only center (loRoCenterMixLevel).

ltRtCenterMixLevel :: Maybe Double

Specify a value for the following Dolby Digital Plus setting: Left total/Right total center mix (Lt/Rt center). MediaConvert uses this value for downmixing. How the service uses this value depends on the value that you choose for Stereo downmix (Eac3StereoDownmix). Valid values: 3.0, 1.5, 0.0, -1.5, -3.0, -4.5, -6.0, and -60. The value -60 mutes the channel. This setting applies only if you keep the default value of 3/2 - L, R, C, Ls, Rs (CODING_MODE_3_2) for the setting Coding mode (Eac3CodingMode). If you choose a different value for Coding mode, the service ignores Left total/Right total center (ltRtCenterMixLevel).

lfeFilter :: Maybe Eac3LfeFilter

Applies a 120Hz lowpass filter to the LFE channel prior to encoding. Only valid with 3_2_LFE coding mode.

dynamicRangeCompressionLine :: Maybe Eac3DynamicRangeCompressionLine

Choose the Dolby Digital dynamic range control (DRC) profile that MediaConvert uses when encoding the metadata in the Dolby Digital stream for the line operating mode. Related setting: When you use this setting, MediaConvert ignores any value you provide for Dynamic range compression profile (DynamicRangeCompressionProfile). For information about the Dolby Digital DRC operating modes and profiles, see the Dynamic Range Control chapter of the Dolby Metadata Guide at https://developer.dolby.com/globalassets/professional/documents/dolby-metadata-guide.pdf.

ltRtSurroundMixLevel :: Maybe Double

Specify a value for the following Dolby Digital Plus setting: Left total/Right total surround mix (Lt/Rt surround). MediaConvert uses this value for downmixing. How the service uses this value depends on the value that you choose for Stereo downmix (Eac3StereoDownmix). Valid values: -1.5, -3.0, -4.5, -6.0, and -60. The value -60 mutes the channel. This setting applies only if you keep the default value of 3/2 - L, R, C, Ls, Rs (CODING_MODE_3_2) for the setting Coding mode (Eac3CodingMode). If you choose a different value for Coding mode, the service ignores Left total/Right total surround (ltRtSurroundMixLevel).

metadataControl :: Maybe Eac3MetadataControl

When set to FOLLOW_INPUT, encoder metadata will be sourced from the DD, DD+, or DolbyE decoder that supplied this audio data. If audio was not supplied from one of these streams, then the static metadata settings will be used.

loRoSurroundMixLevel :: Maybe Double

Specify a value for the following Dolby Digital Plus setting: Left only/Right only (Lo/Ro surround). MediaConvert uses this value for downmixing. How the service uses this value depends on the value that you choose for Stereo downmix (Eac3StereoDownmix). Valid values: -1.5, -3.0, -4.5, -6.0, and -60. The value -60 mutes the channel. This setting applies only if you keep the default value of 3/2 - L, R, C, Ls, Rs (CODING_MODE_3_2) for the setting Coding mode (Eac3CodingMode). If you choose a different value for Coding mode, the service ignores Left only/Right only surround (loRoSurroundMixLevel).

surroundMode :: Maybe Eac3SurroundMode

When encoding 2/0 audio, sets whether Dolby Surround is matrix encoded into the two channels.

attenuationControl :: Maybe Eac3AttenuationControl

If set to ATTENUATE_3_DB, applies a 3 dB attenuation to the surround channels. Only used for 3/2 coding mode.

passthroughControl :: Maybe Eac3PassthroughControl

When set to WHEN_POSSIBLE, input DD+ audio will be passed through if it is present on the input. this detection is dynamic over the life of the transcode. Inputs that alternate between DD+ and non-DD+ content will have a consistent DD+ output as the system alternates between passthrough and encoding.

bitstreamMode :: Maybe Eac3BitstreamMode

Specify the bitstream mode for the E-AC-3 stream that the encoder emits. For more information about the EAC3 bitstream mode, see ATSC A/52-2012 (Annex E).

lfeControl :: Maybe Eac3LfeControl

When encoding 3/2 audio, controls whether the LFE channel is enabled

dynamicRangeCompressionRf :: Maybe Eac3DynamicRangeCompressionRf

Choose the Dolby Digital dynamic range control (DRC) profile that MediaConvert uses when encoding the metadata in the Dolby Digital stream for the RF operating mode. Related setting: When you use this setting, MediaConvert ignores any value you provide for Dynamic range compression profile (DynamicRangeCompressionProfile). For information about the Dolby Digital DRC operating modes and profiles, see the Dynamic Range Control chapter of the Dolby Metadata Guide at https://developer.dolby.com/globalassets/professional/documents/dolby-metadata-guide.pdf.

codingMode :: Maybe Eac3CodingMode

Dolby Digital Plus coding mode. Determines number of channels.

sampleRate :: Maybe Natural

This value is always 48000. It represents the sample rate in Hz.

dcFilter :: Maybe Eac3DcFilter

Activates a DC highpass filter for all input channels.

bitrate :: Maybe Natural

Specify the average bitrate in bits per second. Valid bitrates depend on the coding mode.

phaseControl :: Maybe Eac3PhaseControl

Controls the amount of phase-shift applied to the surround channels. Only used for 3/2 coding mode.

surroundExMode :: Maybe Eac3SurroundExMode

When encoding 3/2 audio, sets whether an extra center back surround channel is matrix encoded into the left and right surround channels.

dialnorm :: Maybe Natural

Sets the dialnorm for the output. If blank and input audio is Dolby Digital Plus, dialnorm will be passed through.

newEac3Settings :: Eac3Settings

Create a value of Eac3Settings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stereoDownmix:Eac3Settings', eac3Settings_stereoDownmix - Choose how the service does stereo downmixing. This setting only applies if you keep the default value of 3/2 - L, R, C, Ls, Rs (CODING_MODE_3_2) for the setting Coding mode (Eac3CodingMode). If you choose a different value for Coding mode, the service ignores Stereo downmix (Eac3StereoDownmix).

$sel:loRoCenterMixLevel:Eac3Settings', eac3Settings_loRoCenterMixLevel - Specify a value for the following Dolby Digital Plus setting: Left only/Right only center mix (Lo/Ro center). MediaConvert uses this value for downmixing. How the service uses this value depends on the value that you choose for Stereo downmix (Eac3StereoDownmix). Valid values: 3.0, 1.5, 0.0, -1.5, -3.0, -4.5, -6.0, and -60. The value -60 mutes the channel. This setting applies only if you keep the default value of 3/2 - L, R, C, Ls, Rs (CODING_MODE_3_2) for the setting Coding mode (Eac3CodingMode). If you choose a different value for Coding mode, the service ignores Left only/Right only center (loRoCenterMixLevel).

$sel:ltRtCenterMixLevel:Eac3Settings', eac3Settings_ltRtCenterMixLevel - Specify a value for the following Dolby Digital Plus setting: Left total/Right total center mix (Lt/Rt center). MediaConvert uses this value for downmixing. How the service uses this value depends on the value that you choose for Stereo downmix (Eac3StereoDownmix). Valid values: 3.0, 1.5, 0.0, -1.5, -3.0, -4.5, -6.0, and -60. The value -60 mutes the channel. This setting applies only if you keep the default value of 3/2 - L, R, C, Ls, Rs (CODING_MODE_3_2) for the setting Coding mode (Eac3CodingMode). If you choose a different value for Coding mode, the service ignores Left total/Right total center (ltRtCenterMixLevel).

$sel:lfeFilter:Eac3Settings', eac3Settings_lfeFilter - Applies a 120Hz lowpass filter to the LFE channel prior to encoding. Only valid with 3_2_LFE coding mode.

$sel:dynamicRangeCompressionLine:Eac3Settings', eac3Settings_dynamicRangeCompressionLine - Choose the Dolby Digital dynamic range control (DRC) profile that MediaConvert uses when encoding the metadata in the Dolby Digital stream for the line operating mode. Related setting: When you use this setting, MediaConvert ignores any value you provide for Dynamic range compression profile (DynamicRangeCompressionProfile). For information about the Dolby Digital DRC operating modes and profiles, see the Dynamic Range Control chapter of the Dolby Metadata Guide at https://developer.dolby.com/globalassets/professional/documents/dolby-metadata-guide.pdf.

$sel:ltRtSurroundMixLevel:Eac3Settings', eac3Settings_ltRtSurroundMixLevel - Specify a value for the following Dolby Digital Plus setting: Left total/Right total surround mix (Lt/Rt surround). MediaConvert uses this value for downmixing. How the service uses this value depends on the value that you choose for Stereo downmix (Eac3StereoDownmix). Valid values: -1.5, -3.0, -4.5, -6.0, and -60. The value -60 mutes the channel. This setting applies only if you keep the default value of 3/2 - L, R, C, Ls, Rs (CODING_MODE_3_2) for the setting Coding mode (Eac3CodingMode). If you choose a different value for Coding mode, the service ignores Left total/Right total surround (ltRtSurroundMixLevel).

$sel:metadataControl:Eac3Settings', eac3Settings_metadataControl - When set to FOLLOW_INPUT, encoder metadata will be sourced from the DD, DD+, or DolbyE decoder that supplied this audio data. If audio was not supplied from one of these streams, then the static metadata settings will be used.

$sel:loRoSurroundMixLevel:Eac3Settings', eac3Settings_loRoSurroundMixLevel - Specify a value for the following Dolby Digital Plus setting: Left only/Right only (Lo/Ro surround). MediaConvert uses this value for downmixing. How the service uses this value depends on the value that you choose for Stereo downmix (Eac3StereoDownmix). Valid values: -1.5, -3.0, -4.5, -6.0, and -60. The value -60 mutes the channel. This setting applies only if you keep the default value of 3/2 - L, R, C, Ls, Rs (CODING_MODE_3_2) for the setting Coding mode (Eac3CodingMode). If you choose a different value for Coding mode, the service ignores Left only/Right only surround (loRoSurroundMixLevel).

$sel:surroundMode:Eac3Settings', eac3Settings_surroundMode - When encoding 2/0 audio, sets whether Dolby Surround is matrix encoded into the two channels.

$sel:attenuationControl:Eac3Settings', eac3Settings_attenuationControl - If set to ATTENUATE_3_DB, applies a 3 dB attenuation to the surround channels. Only used for 3/2 coding mode.

$sel:passthroughControl:Eac3Settings', eac3Settings_passthroughControl - When set to WHEN_POSSIBLE, input DD+ audio will be passed through if it is present on the input. this detection is dynamic over the life of the transcode. Inputs that alternate between DD+ and non-DD+ content will have a consistent DD+ output as the system alternates between passthrough and encoding.

$sel:bitstreamMode:Eac3Settings', eac3Settings_bitstreamMode - Specify the bitstream mode for the E-AC-3 stream that the encoder emits. For more information about the EAC3 bitstream mode, see ATSC A/52-2012 (Annex E).

$sel:lfeControl:Eac3Settings', eac3Settings_lfeControl - When encoding 3/2 audio, controls whether the LFE channel is enabled

$sel:dynamicRangeCompressionRf:Eac3Settings', eac3Settings_dynamicRangeCompressionRf - Choose the Dolby Digital dynamic range control (DRC) profile that MediaConvert uses when encoding the metadata in the Dolby Digital stream for the RF operating mode. Related setting: When you use this setting, MediaConvert ignores any value you provide for Dynamic range compression profile (DynamicRangeCompressionProfile). For information about the Dolby Digital DRC operating modes and profiles, see the Dynamic Range Control chapter of the Dolby Metadata Guide at https://developer.dolby.com/globalassets/professional/documents/dolby-metadata-guide.pdf.

$sel:codingMode:Eac3Settings', eac3Settings_codingMode - Dolby Digital Plus coding mode. Determines number of channels.

$sel:sampleRate:Eac3Settings', eac3Settings_sampleRate - This value is always 48000. It represents the sample rate in Hz.

$sel:dcFilter:Eac3Settings', eac3Settings_dcFilter - Activates a DC highpass filter for all input channels.

$sel:bitrate:Eac3Settings', eac3Settings_bitrate - Specify the average bitrate in bits per second. Valid bitrates depend on the coding mode.

$sel:phaseControl:Eac3Settings', eac3Settings_phaseControl - Controls the amount of phase-shift applied to the surround channels. Only used for 3/2 coding mode.

$sel:surroundExMode:Eac3Settings', eac3Settings_surroundExMode - When encoding 3/2 audio, sets whether an extra center back surround channel is matrix encoded into the left and right surround channels.

$sel:dialnorm:Eac3Settings', eac3Settings_dialnorm - Sets the dialnorm for the output. If blank and input audio is Dolby Digital Plus, dialnorm will be passed through.

eac3Settings_stereoDownmix :: Lens' Eac3Settings (Maybe Eac3StereoDownmix)

Choose how the service does stereo downmixing. This setting only applies if you keep the default value of 3/2 - L, R, C, Ls, Rs (CODING_MODE_3_2) for the setting Coding mode (Eac3CodingMode). If you choose a different value for Coding mode, the service ignores Stereo downmix (Eac3StereoDownmix).

eac3Settings_loRoCenterMixLevel :: Lens' Eac3Settings (Maybe Double)

Specify a value for the following Dolby Digital Plus setting: Left only/Right only center mix (Lo/Ro center). MediaConvert uses this value for downmixing. How the service uses this value depends on the value that you choose for Stereo downmix (Eac3StereoDownmix). Valid values: 3.0, 1.5, 0.0, -1.5, -3.0, -4.5, -6.0, and -60. The value -60 mutes the channel. This setting applies only if you keep the default value of 3/2 - L, R, C, Ls, Rs (CODING_MODE_3_2) for the setting Coding mode (Eac3CodingMode). If you choose a different value for Coding mode, the service ignores Left only/Right only center (loRoCenterMixLevel).

eac3Settings_ltRtCenterMixLevel :: Lens' Eac3Settings (Maybe Double)

Specify a value for the following Dolby Digital Plus setting: Left total/Right total center mix (Lt/Rt center). MediaConvert uses this value for downmixing. How the service uses this value depends on the value that you choose for Stereo downmix (Eac3StereoDownmix). Valid values: 3.0, 1.5, 0.0, -1.5, -3.0, -4.5, -6.0, and -60. The value -60 mutes the channel. This setting applies only if you keep the default value of 3/2 - L, R, C, Ls, Rs (CODING_MODE_3_2) for the setting Coding mode (Eac3CodingMode). If you choose a different value for Coding mode, the service ignores Left total/Right total center (ltRtCenterMixLevel).

eac3Settings_lfeFilter :: Lens' Eac3Settings (Maybe Eac3LfeFilter)

Applies a 120Hz lowpass filter to the LFE channel prior to encoding. Only valid with 3_2_LFE coding mode.

eac3Settings_dynamicRangeCompressionLine :: Lens' Eac3Settings (Maybe Eac3DynamicRangeCompressionLine)

Choose the Dolby Digital dynamic range control (DRC) profile that MediaConvert uses when encoding the metadata in the Dolby Digital stream for the line operating mode. Related setting: When you use this setting, MediaConvert ignores any value you provide for Dynamic range compression profile (DynamicRangeCompressionProfile). For information about the Dolby Digital DRC operating modes and profiles, see the Dynamic Range Control chapter of the Dolby Metadata Guide at https://developer.dolby.com/globalassets/professional/documents/dolby-metadata-guide.pdf.

eac3Settings_ltRtSurroundMixLevel :: Lens' Eac3Settings (Maybe Double)

Specify a value for the following Dolby Digital Plus setting: Left total/Right total surround mix (Lt/Rt surround). MediaConvert uses this value for downmixing. How the service uses this value depends on the value that you choose for Stereo downmix (Eac3StereoDownmix). Valid values: -1.5, -3.0, -4.5, -6.0, and -60. The value -60 mutes the channel. This setting applies only if you keep the default value of 3/2 - L, R, C, Ls, Rs (CODING_MODE_3_2) for the setting Coding mode (Eac3CodingMode). If you choose a different value for Coding mode, the service ignores Left total/Right total surround (ltRtSurroundMixLevel).

eac3Settings_metadataControl :: Lens' Eac3Settings (Maybe Eac3MetadataControl)

When set to FOLLOW_INPUT, encoder metadata will be sourced from the DD, DD+, or DolbyE decoder that supplied this audio data. If audio was not supplied from one of these streams, then the static metadata settings will be used.

eac3Settings_loRoSurroundMixLevel :: Lens' Eac3Settings (Maybe Double)

Specify a value for the following Dolby Digital Plus setting: Left only/Right only (Lo/Ro surround). MediaConvert uses this value for downmixing. How the service uses this value depends on the value that you choose for Stereo downmix (Eac3StereoDownmix). Valid values: -1.5, -3.0, -4.5, -6.0, and -60. The value -60 mutes the channel. This setting applies only if you keep the default value of 3/2 - L, R, C, Ls, Rs (CODING_MODE_3_2) for the setting Coding mode (Eac3CodingMode). If you choose a different value for Coding mode, the service ignores Left only/Right only surround (loRoSurroundMixLevel).

eac3Settings_surroundMode :: Lens' Eac3Settings (Maybe Eac3SurroundMode)

When encoding 2/0 audio, sets whether Dolby Surround is matrix encoded into the two channels.

eac3Settings_attenuationControl :: Lens' Eac3Settings (Maybe Eac3AttenuationControl)

If set to ATTENUATE_3_DB, applies a 3 dB attenuation to the surround channels. Only used for 3/2 coding mode.

eac3Settings_passthroughControl :: Lens' Eac3Settings (Maybe Eac3PassthroughControl)

When set to WHEN_POSSIBLE, input DD+ audio will be passed through if it is present on the input. this detection is dynamic over the life of the transcode. Inputs that alternate between DD+ and non-DD+ content will have a consistent DD+ output as the system alternates between passthrough and encoding.

eac3Settings_bitstreamMode :: Lens' Eac3Settings (Maybe Eac3BitstreamMode)

Specify the bitstream mode for the E-AC-3 stream that the encoder emits. For more information about the EAC3 bitstream mode, see ATSC A/52-2012 (Annex E).

eac3Settings_lfeControl :: Lens' Eac3Settings (Maybe Eac3LfeControl)

When encoding 3/2 audio, controls whether the LFE channel is enabled

eac3Settings_dynamicRangeCompressionRf :: Lens' Eac3Settings (Maybe Eac3DynamicRangeCompressionRf)

Choose the Dolby Digital dynamic range control (DRC) profile that MediaConvert uses when encoding the metadata in the Dolby Digital stream for the RF operating mode. Related setting: When you use this setting, MediaConvert ignores any value you provide for Dynamic range compression profile (DynamicRangeCompressionProfile). For information about the Dolby Digital DRC operating modes and profiles, see the Dynamic Range Control chapter of the Dolby Metadata Guide at https://developer.dolby.com/globalassets/professional/documents/dolby-metadata-guide.pdf.

eac3Settings_codingMode :: Lens' Eac3Settings (Maybe Eac3CodingMode)

Dolby Digital Plus coding mode. Determines number of channels.

eac3Settings_sampleRate :: Lens' Eac3Settings (Maybe Natural)

This value is always 48000. It represents the sample rate in Hz.

eac3Settings_dcFilter :: Lens' Eac3Settings (Maybe Eac3DcFilter)

Activates a DC highpass filter for all input channels.

eac3Settings_bitrate :: Lens' Eac3Settings (Maybe Natural)

Specify the average bitrate in bits per second. Valid bitrates depend on the coding mode.

eac3Settings_phaseControl :: Lens' Eac3Settings (Maybe Eac3PhaseControl)

Controls the amount of phase-shift applied to the surround channels. Only used for 3/2 coding mode.

eac3Settings_surroundExMode :: Lens' Eac3Settings (Maybe Eac3SurroundExMode)

When encoding 3/2 audio, sets whether an extra center back surround channel is matrix encoded into the left and right surround channels.

eac3Settings_dialnorm :: Lens' Eac3Settings (Maybe Natural)

Sets the dialnorm for the output. If blank and input audio is Dolby Digital Plus, dialnorm will be passed through.