Network.AWS.MediaConvert.Types.FileSourceSettings

Description

Documentation

data FileSourceSettings

If your input captions are SCC, SMI, SRT, STL, TTML, WebVTT, or IMSC 1.1 in an xml file, specify the URI of the input caption source file. If your caption source is IMSC in an IMF package, use TrackSourceSettings instead of FileSoureSettings.

See: newFileSourceSettings smart constructor.

Constructors

FileSourceSettings'

Fields

convert608To708 :: Maybe FileSourceConvert608To708

Specify whether this set of input captions appears in your outputs in both 608 and 708 format. If you choose Upconvert (UPCONVERT), MediaConvert includes the captions data in two ways: it passes the 608 data through using the 608 compatibility bytes fields of the 708 wrapper, and it also translates the 608 data into 708.

framerate :: Maybe CaptionSourceFramerate

Ignore this setting unless your input captions format is SCC. To have the service compensate for differing frame rates between your input captions and input video, specify the frame rate of the captions file. Specify this value as a fraction, using the settings Framerate numerator (framerateNumerator) and Framerate denominator (framerateDenominator). For example, you might specify 24 / 1 for 24 fps, 25 / 1 for 25 fps, 24000 / 1001 for 23.976 fps, or 30000 / 1001 for 29.97 fps.

sourceFile :: Maybe Text

External caption file used for loading captions. Accepted file extensions are 'scc', 'ttml', 'dfxp', 'stl', 'srt', 'xml', 'smi', 'webvtt', and 'vtt'.

timeDelta :: Maybe Int

Specifies a time delta in seconds to offset the captions from the source file.

newFileSourceSettings :: FileSourceSettings

Create a value of FileSourceSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:convert608To708:FileSourceSettings', fileSourceSettings_convert608To708 - Specify whether this set of input captions appears in your outputs in both 608 and 708 format. If you choose Upconvert (UPCONVERT), MediaConvert includes the captions data in two ways: it passes the 608 data through using the 608 compatibility bytes fields of the 708 wrapper, and it also translates the 608 data into 708.

$sel:framerate:FileSourceSettings', fileSourceSettings_framerate - Ignore this setting unless your input captions format is SCC. To have the service compensate for differing frame rates between your input captions and input video, specify the frame rate of the captions file. Specify this value as a fraction, using the settings Framerate numerator (framerateNumerator) and Framerate denominator (framerateDenominator). For example, you might specify 24 / 1 for 24 fps, 25 / 1 for 25 fps, 24000 / 1001 for 23.976 fps, or 30000 / 1001 for 29.97 fps.

$sel:sourceFile:FileSourceSettings', fileSourceSettings_sourceFile - External caption file used for loading captions. Accepted file extensions are 'scc', 'ttml', 'dfxp', 'stl', 'srt', 'xml', 'smi', 'webvtt', and 'vtt'.

$sel:timeDelta:FileSourceSettings', fileSourceSettings_timeDelta - Specifies a time delta in seconds to offset the captions from the source file.

fileSourceSettings_convert608To708 :: Lens' FileSourceSettings (Maybe FileSourceConvert608To708)

Specify whether this set of input captions appears in your outputs in both 608 and 708 format. If you choose Upconvert (UPCONVERT), MediaConvert includes the captions data in two ways: it passes the 608 data through using the 608 compatibility bytes fields of the 708 wrapper, and it also translates the 608 data into 708.

fileSourceSettings_framerate :: Lens' FileSourceSettings (Maybe CaptionSourceFramerate)

Ignore this setting unless your input captions format is SCC. To have the service compensate for differing frame rates between your input captions and input video, specify the frame rate of the captions file. Specify this value as a fraction, using the settings Framerate numerator (framerateNumerator) and Framerate denominator (framerateDenominator). For example, you might specify 24 / 1 for 24 fps, 25 / 1 for 25 fps, 24000 / 1001 for 23.976 fps, or 30000 / 1001 for 29.97 fps.

fileSourceSettings_sourceFile :: Lens' FileSourceSettings (Maybe Text)

External caption file used for loading captions. Accepted file extensions are 'scc', 'ttml', 'dfxp', 'stl', 'srt', 'xml', 'smi', 'webvtt', and 'vtt'.

fileSourceSettings_timeDelta :: Lens' FileSourceSettings (Maybe Int)

Specifies a time delta in seconds to offset the captions from the source file.