Network.AWS.MediaConvert.Types.Hdr10Metadata

Description

Documentation

data Hdr10Metadata

Use these settings to specify static color calibration metadata, as defined by SMPTE ST 2086. These values don't affect the pixel values that are encoded in the video stream. They are intended to help the downstream video player display content in a way that reflects the intentions of the the content creator.

See: newHdr10Metadata smart constructor.

Constructors

Hdr10Metadata'

Fields

redPrimaryX :: Maybe Natural

HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

bluePrimaryX :: Maybe Natural

HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

maxFrameAverageLightLevel :: Maybe Natural

Maximum average light level of any frame in the coded video sequence, in units of candelas per square meter. This setting doesn't have a default value; you must specify a value that is suitable for the content.

whitePointY :: Maybe Natural

HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

maxContentLightLevel :: Maybe Natural

Maximum light level among all samples in the coded video sequence, in units of candelas per square meter. This setting doesn't have a default value; you must specify a value that is suitable for the content.

whitePointX :: Maybe Natural

HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

bluePrimaryY :: Maybe Natural

HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

greenPrimaryY :: Maybe Natural

HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

greenPrimaryX :: Maybe Natural

HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

minLuminance :: Maybe Natural

Nominal minimum mastering display luminance in units of of 0.0001 candelas per square meter

redPrimaryY :: Maybe Natural

HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

maxLuminance :: Maybe Natural

Nominal maximum mastering display luminance in units of of 0.0001 candelas per square meter.

newHdr10Metadata :: Hdr10Metadata

Create a value of Hdr10Metadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:redPrimaryX:Hdr10Metadata', hdr10Metadata_redPrimaryX - HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

$sel:bluePrimaryX:Hdr10Metadata', hdr10Metadata_bluePrimaryX - HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

$sel:maxFrameAverageLightLevel:Hdr10Metadata', hdr10Metadata_maxFrameAverageLightLevel - Maximum average light level of any frame in the coded video sequence, in units of candelas per square meter. This setting doesn't have a default value; you must specify a value that is suitable for the content.

$sel:whitePointY:Hdr10Metadata', hdr10Metadata_whitePointY - HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

$sel:maxContentLightLevel:Hdr10Metadata', hdr10Metadata_maxContentLightLevel - Maximum light level among all samples in the coded video sequence, in units of candelas per square meter. This setting doesn't have a default value; you must specify a value that is suitable for the content.

$sel:whitePointX:Hdr10Metadata', hdr10Metadata_whitePointX - HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

$sel:bluePrimaryY:Hdr10Metadata', hdr10Metadata_bluePrimaryY - HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

$sel:greenPrimaryY:Hdr10Metadata', hdr10Metadata_greenPrimaryY - HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

$sel:greenPrimaryX:Hdr10Metadata', hdr10Metadata_greenPrimaryX - HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

$sel:minLuminance:Hdr10Metadata', hdr10Metadata_minLuminance - Nominal minimum mastering display luminance in units of of 0.0001 candelas per square meter

$sel:redPrimaryY:Hdr10Metadata', hdr10Metadata_redPrimaryY - HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

$sel:maxLuminance:Hdr10Metadata', hdr10Metadata_maxLuminance - Nominal maximum mastering display luminance in units of of 0.0001 candelas per square meter.

hdr10Metadata_redPrimaryX :: Lens' Hdr10Metadata (Maybe Natural)

HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

hdr10Metadata_bluePrimaryX :: Lens' Hdr10Metadata (Maybe Natural)

HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

hdr10Metadata_maxFrameAverageLightLevel :: Lens' Hdr10Metadata (Maybe Natural)

Maximum average light level of any frame in the coded video sequence, in units of candelas per square meter. This setting doesn't have a default value; you must specify a value that is suitable for the content.

hdr10Metadata_whitePointY :: Lens' Hdr10Metadata (Maybe Natural)

HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

hdr10Metadata_maxContentLightLevel :: Lens' Hdr10Metadata (Maybe Natural)

Maximum light level among all samples in the coded video sequence, in units of candelas per square meter. This setting doesn't have a default value; you must specify a value that is suitable for the content.

hdr10Metadata_whitePointX :: Lens' Hdr10Metadata (Maybe Natural)

HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

hdr10Metadata_bluePrimaryY :: Lens' Hdr10Metadata (Maybe Natural)

HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

hdr10Metadata_greenPrimaryY :: Lens' Hdr10Metadata (Maybe Natural)

HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

hdr10Metadata_greenPrimaryX :: Lens' Hdr10Metadata (Maybe Natural)

HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

hdr10Metadata_minLuminance :: Lens' Hdr10Metadata (Maybe Natural)

Nominal minimum mastering display luminance in units of of 0.0001 candelas per square meter

hdr10Metadata_redPrimaryY :: Lens' Hdr10Metadata (Maybe Natural)

HDR Master Display Information must be provided by a color grader, using color grading tools. Range is 0 to 50,000, each increment represents 0.00002 in CIE1931 color coordinate. Note that this setting is not for color correction.

hdr10Metadata_maxLuminance :: Lens' Hdr10Metadata (Maybe Natural)

Nominal maximum mastering display luminance in units of of 0.0001 candelas per square meter.