Network.AWS.MediaConvert.Types.HopDestination

Description

Documentation

data HopDestination

Optional. Configuration for a destination queue to which the job can hop once a customer-defined minimum wait time has passed.

See: newHopDestination smart constructor.

Constructors

HopDestination'

Fields

priority :: Maybe Int

Optional. When you set up a job to use queue hopping, you can specify a different relative priority for the job in the destination queue. If you don't specify, the relative priority will remain the same as in the previous queue.

queue :: Maybe Text

Optional unless the job is submitted on the default queue. When you set up a job to use queue hopping, you can specify a destination queue. This queue cannot be the original queue to which the job is submitted. If the original queue isn't the default queue and you don't specify the destination queue, the job will move to the default queue.

waitMinutes :: Maybe Int

Required for setting up a job to use queue hopping. Minimum wait time in minutes until the job can hop to the destination queue. Valid range is 1 to 1440 minutes, inclusive.

newHopDestination :: HopDestination

Create a value of HopDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:priority:HopDestination', hopDestination_priority - Optional. When you set up a job to use queue hopping, you can specify a different relative priority for the job in the destination queue. If you don't specify, the relative priority will remain the same as in the previous queue.

$sel:queue:HopDestination', hopDestination_queue - Optional unless the job is submitted on the default queue. When you set up a job to use queue hopping, you can specify a destination queue. This queue cannot be the original queue to which the job is submitted. If the original queue isn't the default queue and you don't specify the destination queue, the job will move to the default queue.

$sel:waitMinutes:HopDestination', hopDestination_waitMinutes - Required for setting up a job to use queue hopping. Minimum wait time in minutes until the job can hop to the destination queue. Valid range is 1 to 1440 minutes, inclusive.

hopDestination_priority :: Lens' HopDestination (Maybe Int)

Optional. When you set up a job to use queue hopping, you can specify a different relative priority for the job in the destination queue. If you don't specify, the relative priority will remain the same as in the previous queue.

hopDestination_queue :: Lens' HopDestination (Maybe Text)

Optional unless the job is submitted on the default queue. When you set up a job to use queue hopping, you can specify a destination queue. This queue cannot be the original queue to which the job is submitted. If the original queue isn't the default queue and you don't specify the destination queue, the job will move to the default queue.

hopDestination_waitMinutes :: Lens' HopDestination (Maybe Int)

Required for setting up a job to use queue hopping. Minimum wait time in minutes until the job can hop to the destination queue. Valid range is 1 to 1440 minutes, inclusive.