Network.AWS.MediaConvert.Types.InputDecryptionSettings

Description

Documentation

data InputDecryptionSettings

Settings for decrypting any input files that you encrypt before you upload them to Amazon S3. MediaConvert can decrypt files only when you use AWS Key Management Service (KMS) to encrypt the data key that you use to encrypt your content.

See: newInputDecryptionSettings smart constructor.

Constructors

InputDecryptionSettings'

Fields

decryptionMode :: Maybe DecryptionMode

Specify the encryption mode that you used to encrypt your input files.

encryptedDecryptionKey :: Maybe Text

Warning! Don't provide your encryption key in plaintext. Your job settings could be intercepted, making your encrypted content vulnerable. Specify the encrypted version of the data key that you used to encrypt your content. The data key must be encrypted by AWS Key Management Service (KMS). The key can be 128, 192, or 256 bits.

initializationVector :: Maybe Text

Specify the initialization vector that you used when you encrypted your content before uploading it to Amazon S3. You can use a 16-byte initialization vector with any encryption mode. Or, you can use a 12-byte initialization vector with GCM or CTR. MediaConvert accepts only initialization vectors that are base64-encoded.

kmsKeyRegion :: Maybe Text

Specify the AWS Region for AWS Key Management Service (KMS) that you used to encrypt your data key, if that Region is different from the one you are using for AWS Elemental MediaConvert.

newInputDecryptionSettings :: InputDecryptionSettings

Create a value of InputDecryptionSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:decryptionMode:InputDecryptionSettings', inputDecryptionSettings_decryptionMode - Specify the encryption mode that you used to encrypt your input files.

$sel:encryptedDecryptionKey:InputDecryptionSettings', inputDecryptionSettings_encryptedDecryptionKey - Warning! Don't provide your encryption key in plaintext. Your job settings could be intercepted, making your encrypted content vulnerable. Specify the encrypted version of the data key that you used to encrypt your content. The data key must be encrypted by AWS Key Management Service (KMS). The key can be 128, 192, or 256 bits.

$sel:initializationVector:InputDecryptionSettings', inputDecryptionSettings_initializationVector - Specify the initialization vector that you used when you encrypted your content before uploading it to Amazon S3. You can use a 16-byte initialization vector with any encryption mode. Or, you can use a 12-byte initialization vector with GCM or CTR. MediaConvert accepts only initialization vectors that are base64-encoded.

$sel:kmsKeyRegion:InputDecryptionSettings', inputDecryptionSettings_kmsKeyRegion - Specify the AWS Region for AWS Key Management Service (KMS) that you used to encrypt your data key, if that Region is different from the one you are using for AWS Elemental MediaConvert.

inputDecryptionSettings_decryptionMode :: Lens' InputDecryptionSettings (Maybe DecryptionMode)

Specify the encryption mode that you used to encrypt your input files.

inputDecryptionSettings_encryptedDecryptionKey :: Lens' InputDecryptionSettings (Maybe Text)

Warning! Don't provide your encryption key in plaintext. Your job settings could be intercepted, making your encrypted content vulnerable. Specify the encrypted version of the data key that you used to encrypt your content. The data key must be encrypted by AWS Key Management Service (KMS). The key can be 128, 192, or 256 bits.

inputDecryptionSettings_initializationVector :: Lens' InputDecryptionSettings (Maybe Text)

Specify the initialization vector that you used when you encrypted your content before uploading it to Amazon S3. You can use a 16-byte initialization vector with any encryption mode. Or, you can use a 12-byte initialization vector with GCM or CTR. MediaConvert accepts only initialization vectors that are base64-encoded.

inputDecryptionSettings_kmsKeyRegion :: Lens' InputDecryptionSettings (Maybe Text)

Specify the AWS Region for AWS Key Management Service (KMS) that you used to encrypt your data key, if that Region is different from the one you are using for AWS Elemental MediaConvert.