Network.AWS.MediaConvert.Types.InsertableImage

Description

Documentation

data InsertableImage

These settings apply to a specific graphic overlay. You can include multiple overlays in your job.

See: newInsertableImage smart constructor.

Constructors

InsertableImage'

Fields

imageX :: Maybe Natural

Specify the distance, in pixels, between the inserted image and the left edge of the video frame. Required for any image overlay that you specify.

height :: Maybe Natural

Specify the height of the inserted image in pixels. If you specify a value that's larger than the video resolution height, the service will crop your overlaid image to fit. To use the native height of the image, keep this setting blank.

startTime :: Maybe Text

Specify the timecode of the frame that you want the overlay to first appear on. This must be in timecode (HH:MM:SS:FF or HH:MM:SS;FF) format. Remember to take into account your timecode source settings.

fadeOut :: Maybe Natural

Specify the length of time, in milliseconds, between the end of the time that you have specified for the image overlay Duration and when the overlaid image has faded to total transparency. If you don't specify a value for Fade-out, the image will disappear abruptly at the end of the inserted image duration.

width :: Maybe Natural

Specify the width of the inserted image in pixels. If you specify a value that's larger than the video resolution width, the service will crop your overlaid image to fit. To use the native width of the image, keep this setting blank.

opacity :: Maybe Natural

Use Opacity (Opacity) to specify how much of the underlying video shows through the inserted image. 0 is transparent and 100 is fully opaque. Default is 50.

layer :: Maybe Natural

Specify how overlapping inserted images appear. Images with higher values for Layer appear on top of images with lower values for Layer.

duration :: Maybe Natural

Specify the time, in milliseconds, for the image to remain on the output video. This duration includes fade-in time but not fade-out time.

imageY :: Maybe Natural

Specify the distance, in pixels, between the overlaid image and the top edge of the video frame. Required for any image overlay that you specify.

imageInserterInput :: Maybe Text

Specify the HTTP, HTTPS, or Amazon S3 location of the image that you want to overlay on the video. Use a PNG or TGA file.

fadeIn :: Maybe Natural

Specify the length of time, in milliseconds, between the Start time that you specify for the image insertion and the time that the image appears at full opacity. Full opacity is the level that you specify for the opacity setting. If you don't specify a value for Fade-in, the image will appear abruptly at the overlay start time.

newInsertableImage :: InsertableImage

Create a value of InsertableImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageX:InsertableImage', insertableImage_imageX - Specify the distance, in pixels, between the inserted image and the left edge of the video frame. Required for any image overlay that you specify.

$sel:height:InsertableImage', insertableImage_height - Specify the height of the inserted image in pixels. If you specify a value that's larger than the video resolution height, the service will crop your overlaid image to fit. To use the native height of the image, keep this setting blank.

$sel:startTime:InsertableImage', insertableImage_startTime - Specify the timecode of the frame that you want the overlay to first appear on. This must be in timecode (HH:MM:SS:FF or HH:MM:SS;FF) format. Remember to take into account your timecode source settings.

$sel:fadeOut:InsertableImage', insertableImage_fadeOut - Specify the length of time, in milliseconds, between the end of the time that you have specified for the image overlay Duration and when the overlaid image has faded to total transparency. If you don't specify a value for Fade-out, the image will disappear abruptly at the end of the inserted image duration.

$sel:width:InsertableImage', insertableImage_width - Specify the width of the inserted image in pixels. If you specify a value that's larger than the video resolution width, the service will crop your overlaid image to fit. To use the native width of the image, keep this setting blank.

$sel:opacity:InsertableImage', insertableImage_opacity - Use Opacity (Opacity) to specify how much of the underlying video shows through the inserted image. 0 is transparent and 100 is fully opaque. Default is 50.

$sel:layer:InsertableImage', insertableImage_layer - Specify how overlapping inserted images appear. Images with higher values for Layer appear on top of images with lower values for Layer.

$sel:duration:InsertableImage', insertableImage_duration - Specify the time, in milliseconds, for the image to remain on the output video. This duration includes fade-in time but not fade-out time.

$sel:imageY:InsertableImage', insertableImage_imageY - Specify the distance, in pixels, between the overlaid image and the top edge of the video frame. Required for any image overlay that you specify.

$sel:imageInserterInput:InsertableImage', insertableImage_imageInserterInput - Specify the HTTP, HTTPS, or Amazon S3 location of the image that you want to overlay on the video. Use a PNG or TGA file.

$sel:fadeIn:InsertableImage', insertableImage_fadeIn - Specify the length of time, in milliseconds, between the Start time that you specify for the image insertion and the time that the image appears at full opacity. Full opacity is the level that you specify for the opacity setting. If you don't specify a value for Fade-in, the image will appear abruptly at the overlay start time.

insertableImage_imageX :: Lens' InsertableImage (Maybe Natural)

Specify the distance, in pixels, between the inserted image and the left edge of the video frame. Required for any image overlay that you specify.

insertableImage_height :: Lens' InsertableImage (Maybe Natural)

Specify the height of the inserted image in pixels. If you specify a value that's larger than the video resolution height, the service will crop your overlaid image to fit. To use the native height of the image, keep this setting blank.

insertableImage_startTime :: Lens' InsertableImage (Maybe Text)

Specify the timecode of the frame that you want the overlay to first appear on. This must be in timecode (HH:MM:SS:FF or HH:MM:SS;FF) format. Remember to take into account your timecode source settings.

insertableImage_fadeOut :: Lens' InsertableImage (Maybe Natural)

Specify the length of time, in milliseconds, between the end of the time that you have specified for the image overlay Duration and when the overlaid image has faded to total transparency. If you don't specify a value for Fade-out, the image will disappear abruptly at the end of the inserted image duration.

insertableImage_width :: Lens' InsertableImage (Maybe Natural)

Specify the width of the inserted image in pixels. If you specify a value that's larger than the video resolution width, the service will crop your overlaid image to fit. To use the native width of the image, keep this setting blank.

insertableImage_opacity :: Lens' InsertableImage (Maybe Natural)

Use Opacity (Opacity) to specify how much of the underlying video shows through the inserted image. 0 is transparent and 100 is fully opaque. Default is 50.

insertableImage_layer :: Lens' InsertableImage (Maybe Natural)

Specify how overlapping inserted images appear. Images with higher values for Layer appear on top of images with lower values for Layer.

insertableImage_duration :: Lens' InsertableImage (Maybe Natural)

Specify the time, in milliseconds, for the image to remain on the output video. This duration includes fade-in time but not fade-out time.

insertableImage_imageY :: Lens' InsertableImage (Maybe Natural)

Specify the distance, in pixels, between the overlaid image and the top edge of the video frame. Required for any image overlay that you specify.

insertableImage_imageInserterInput :: Lens' InsertableImage (Maybe Text)

Specify the HTTP, HTTPS, or Amazon S3 location of the image that you want to overlay on the video. Use a PNG or TGA file.

insertableImage_fadeIn :: Lens' InsertableImage (Maybe Natural)

Specify the length of time, in milliseconds, between the Start time that you specify for the image insertion and the time that the image appears at full opacity. Full opacity is the level that you specify for the opacity setting. If you don't specify a value for Fade-in, the image will appear abruptly at the overlay start time.