Network.AWS.MediaConvert.Types.Job

Description

Documentation

data Job

Each job converts an input file into an output file or files. For more information, see the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html

See: newJob smart constructor.

Constructors

Job'

Fields

billingTagsSource :: Maybe BillingTagsSource

The tag type that AWS Billing and Cost Management will use to sort your AWS Elemental MediaConvert costs on any billing report that you set up.

accelerationSettings :: Maybe AccelerationSettings

Accelerated transcoding can significantly speed up jobs with long, visually complex content.

status :: Maybe JobStatus

A job's status can be SUBMITTED, PROGRESSING, COMPLETE, CANCELED, or ERROR.

accelerationStatus :: Maybe AccelerationStatus

Describes whether the current job is running with accelerated transcoding. For jobs that have Acceleration (AccelerationMode) set to DISABLED, AccelerationStatus is always NOT_APPLICABLE. For jobs that have Acceleration (AccelerationMode) set to ENABLED or PREFERRED, AccelerationStatus is one of the other states. AccelerationStatus is IN_PROGRESS initially, while the service determines whether the input files and job settings are compatible with accelerated transcoding. If they are, AcclerationStatus is ACCELERATED. If your input files and job settings aren't compatible with accelerated transcoding, the service either fails your job or runs it without accelerated transcoding, depending on how you set Acceleration (AccelerationMode). When the service runs your job without accelerated transcoding, AccelerationStatus is NOT_ACCELERATED.

retryCount :: Maybe Int

The number of times that the service automatically attempted to process your job after encountering an error.

createdAt :: Maybe POSIX

The time, in Unix epoch format in seconds, when the job got created.

queueTransitions :: Maybe [QueueTransition]

The job's queue hopping history.

arn :: Maybe Text

An identifier for this resource that is unique within all of AWS.

id :: Maybe Text

A portion of the job's ARN, unique within your AWS Elemental MediaConvert resources

jobPercentComplete :: Maybe Int

An estimate of how far your job has progressed. This estimate is shown as a percentage of the total time from when your job leaves its queue to when your output files appear in your output Amazon S3 bucket. AWS Elemental MediaConvert provides jobPercentComplete in CloudWatch STATUS_UPDATE events and in the response to GetJob and ListJobs requests. The jobPercentComplete estimate is reliable for the following input containers: Quicktime, Transport Stream, MP4, and MXF. For some jobs, the service can't provide information about job progress. In those cases, jobPercentComplete returns a null value.

statusUpdateInterval :: Maybe StatusUpdateInterval

Specify how often MediaConvert sends STATUS_UPDATE events to Amazon CloudWatch Events. Set the interval, in seconds, between status updates. MediaConvert sends an update at this interval from the time the service begins processing your job to the time it completes the transcode or encounters an error.

priority :: Maybe Int

Relative priority on the job.

jobTemplate :: Maybe Text

The job template that the job is created from, if it is created from a job template.

userMetadata :: Maybe (HashMap Text Text)

User-defined metadata that you want to associate with an MediaConvert job. You specify metadata in key/value pairs.

queue :: Maybe Text

When you create a job, you can specify a queue to send it to. If you don't specify, the job will go to the default queue. For more about queues, see the User Guide topic at https://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html

currentPhase :: Maybe JobPhase

A job's phase can be PROBING, TRANSCODING OR UPLOADING

errorMessage :: Maybe Text

Error message of Job

simulateReservedQueue :: Maybe SimulateReservedQueue

Enable this setting when you run a test job to estimate how many reserved transcoding slots (RTS) you need. When this is enabled, MediaConvert runs your job from an on-demand queue with similar performance to what you will see with one RTS in a reserved queue. This setting is disabled by default.

timing :: Maybe Timing

Information about when jobs are submitted, started, and finished is specified in Unix epoch format in seconds.

hopDestinations :: Maybe [HopDestination]

Optional list of hop destinations.

messages :: Maybe JobMessages

Provides messages from the service about jobs that you have already successfully submitted.

errorCode :: Maybe Int

Error code for the job

outputGroupDetails :: Maybe [OutputGroupDetail]

List of output group details

role' :: Text

The IAM role you use for creating this job. For details about permissions, see the User Guide topic at the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/iam-role.html

settings :: JobSettings

JobSettings contains all the transcode settings for a job.

newJob

Arguments

:: Text
$sel:role':Job'

-> JobSettings
$sel:settings:Job'

-> Job

Create a value of Job with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:billingTagsSource:Job', job_billingTagsSource - The tag type that AWS Billing and Cost Management will use to sort your AWS Elemental MediaConvert costs on any billing report that you set up.

$sel:accelerationSettings:Job', job_accelerationSettings - Accelerated transcoding can significantly speed up jobs with long, visually complex content.

$sel:status:Job', job_status - A job's status can be SUBMITTED, PROGRESSING, COMPLETE, CANCELED, or ERROR.

$sel:accelerationStatus:Job', job_accelerationStatus - Describes whether the current job is running with accelerated transcoding. For jobs that have Acceleration (AccelerationMode) set to DISABLED, AccelerationStatus is always NOT_APPLICABLE. For jobs that have Acceleration (AccelerationMode) set to ENABLED or PREFERRED, AccelerationStatus is one of the other states. AccelerationStatus is IN_PROGRESS initially, while the service determines whether the input files and job settings are compatible with accelerated transcoding. If they are, AcclerationStatus is ACCELERATED. If your input files and job settings aren't compatible with accelerated transcoding, the service either fails your job or runs it without accelerated transcoding, depending on how you set Acceleration (AccelerationMode). When the service runs your job without accelerated transcoding, AccelerationStatus is NOT_ACCELERATED.

$sel:retryCount:Job', job_retryCount - The number of times that the service automatically attempted to process your job after encountering an error.

$sel:createdAt:Job', job_createdAt - The time, in Unix epoch format in seconds, when the job got created.

$sel:queueTransitions:Job', job_queueTransitions - The job's queue hopping history.

$sel:arn:Job', job_arn - An identifier for this resource that is unique within all of AWS.

$sel:id:Job', job_id - A portion of the job's ARN, unique within your AWS Elemental MediaConvert resources

$sel:jobPercentComplete:Job', job_jobPercentComplete - An estimate of how far your job has progressed. This estimate is shown as a percentage of the total time from when your job leaves its queue to when your output files appear in your output Amazon S3 bucket. AWS Elemental MediaConvert provides jobPercentComplete in CloudWatch STATUS_UPDATE events and in the response to GetJob and ListJobs requests. The jobPercentComplete estimate is reliable for the following input containers: Quicktime, Transport Stream, MP4, and MXF. For some jobs, the service can't provide information about job progress. In those cases, jobPercentComplete returns a null value.

$sel:statusUpdateInterval:Job', job_statusUpdateInterval - Specify how often MediaConvert sends STATUS_UPDATE events to Amazon CloudWatch Events. Set the interval, in seconds, between status updates. MediaConvert sends an update at this interval from the time the service begins processing your job to the time it completes the transcode or encounters an error.

$sel:priority:Job', job_priority - Relative priority on the job.

$sel:jobTemplate:Job', job_jobTemplate - The job template that the job is created from, if it is created from a job template.

$sel:userMetadata:Job', job_userMetadata - User-defined metadata that you want to associate with an MediaConvert job. You specify metadata in key/value pairs.

$sel:queue:Job', job_queue - When you create a job, you can specify a queue to send it to. If you don't specify, the job will go to the default queue. For more about queues, see the User Guide topic at https://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html

$sel:currentPhase:Job', job_currentPhase - A job's phase can be PROBING, TRANSCODING OR UPLOADING

$sel:errorMessage:Job', job_errorMessage - Error message of Job

$sel:simulateReservedQueue:Job', job_simulateReservedQueue - Enable this setting when you run a test job to estimate how many reserved transcoding slots (RTS) you need. When this is enabled, MediaConvert runs your job from an on-demand queue with similar performance to what you will see with one RTS in a reserved queue. This setting is disabled by default.

$sel:timing:Job', job_timing - Information about when jobs are submitted, started, and finished is specified in Unix epoch format in seconds.

$sel:hopDestinations:Job', job_hopDestinations - Optional list of hop destinations.

$sel:messages:Job', job_messages - Provides messages from the service about jobs that you have already successfully submitted.

$sel:errorCode:Job', job_errorCode - Error code for the job

$sel:outputGroupDetails:Job', job_outputGroupDetails - List of output group details

$sel:role':Job', job_role - The IAM role you use for creating this job. For details about permissions, see the User Guide topic at the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/iam-role.html

$sel:settings:Job', job_settings - JobSettings contains all the transcode settings for a job.

job_billingTagsSource :: Lens' Job (Maybe BillingTagsSource)

The tag type that AWS Billing and Cost Management will use to sort your AWS Elemental MediaConvert costs on any billing report that you set up.

job_accelerationSettings :: Lens' Job (Maybe AccelerationSettings)

Accelerated transcoding can significantly speed up jobs with long, visually complex content.

job_status :: Lens' Job (Maybe JobStatus)

A job's status can be SUBMITTED, PROGRESSING, COMPLETE, CANCELED, or ERROR.

job_accelerationStatus :: Lens' Job (Maybe AccelerationStatus)

Describes whether the current job is running with accelerated transcoding. For jobs that have Acceleration (AccelerationMode) set to DISABLED, AccelerationStatus is always NOT_APPLICABLE. For jobs that have Acceleration (AccelerationMode) set to ENABLED or PREFERRED, AccelerationStatus is one of the other states. AccelerationStatus is IN_PROGRESS initially, while the service determines whether the input files and job settings are compatible with accelerated transcoding. If they are, AcclerationStatus is ACCELERATED. If your input files and job settings aren't compatible with accelerated transcoding, the service either fails your job or runs it without accelerated transcoding, depending on how you set Acceleration (AccelerationMode). When the service runs your job without accelerated transcoding, AccelerationStatus is NOT_ACCELERATED.

job_retryCount :: Lens' Job (Maybe Int)

The number of times that the service automatically attempted to process your job after encountering an error.

job_createdAt :: Lens' Job (Maybe UTCTime)

The time, in Unix epoch format in seconds, when the job got created.

job_queueTransitions :: Lens' Job (Maybe [QueueTransition])

The job's queue hopping history.

job_arn :: Lens' Job (Maybe Text)

An identifier for this resource that is unique within all of AWS.

job_id :: Lens' Job (Maybe Text)

A portion of the job's ARN, unique within your AWS Elemental MediaConvert resources

job_jobPercentComplete :: Lens' Job (Maybe Int)

An estimate of how far your job has progressed. This estimate is shown as a percentage of the total time from when your job leaves its queue to when your output files appear in your output Amazon S3 bucket. AWS Elemental MediaConvert provides jobPercentComplete in CloudWatch STATUS_UPDATE events and in the response to GetJob and ListJobs requests. The jobPercentComplete estimate is reliable for the following input containers: Quicktime, Transport Stream, MP4, and MXF. For some jobs, the service can't provide information about job progress. In those cases, jobPercentComplete returns a null value.

job_statusUpdateInterval :: Lens' Job (Maybe StatusUpdateInterval)

Specify how often MediaConvert sends STATUS_UPDATE events to Amazon CloudWatch Events. Set the interval, in seconds, between status updates. MediaConvert sends an update at this interval from the time the service begins processing your job to the time it completes the transcode or encounters an error.

job_priority :: Lens' Job (Maybe Int)

Relative priority on the job.

job_jobTemplate :: Lens' Job (Maybe Text)

The job template that the job is created from, if it is created from a job template.

job_userMetadata :: Lens' Job (Maybe (HashMap Text Text))

User-defined metadata that you want to associate with an MediaConvert job. You specify metadata in key/value pairs.

job_queue :: Lens' Job (Maybe Text)

When you create a job, you can specify a queue to send it to. If you don't specify, the job will go to the default queue. For more about queues, see the User Guide topic at https://docs.aws.amazon.com/mediaconvert/latest/ug/what-is.html

job_currentPhase :: Lens' Job (Maybe JobPhase)

A job's phase can be PROBING, TRANSCODING OR UPLOADING

job_errorMessage :: Lens' Job (Maybe Text)

Error message of Job

job_simulateReservedQueue :: Lens' Job (Maybe SimulateReservedQueue)

Enable this setting when you run a test job to estimate how many reserved transcoding slots (RTS) you need. When this is enabled, MediaConvert runs your job from an on-demand queue with similar performance to what you will see with one RTS in a reserved queue. This setting is disabled by default.

job_timing :: Lens' Job (Maybe Timing)

Information about when jobs are submitted, started, and finished is specified in Unix epoch format in seconds.

job_hopDestinations :: Lens' Job (Maybe [HopDestination])

Optional list of hop destinations.

job_messages :: Lens' Job (Maybe JobMessages)

Provides messages from the service about jobs that you have already successfully submitted.

job_errorCode :: Lens' Job (Maybe Int)

Error code for the job

job_outputGroupDetails :: Lens' Job (Maybe [OutputGroupDetail])

List of output group details

job_role :: Lens' Job Text

The IAM role you use for creating this job. For details about permissions, see the User Guide topic at the User Guide at https://docs.aws.amazon.com/mediaconvert/latest/ug/iam-role.html

job_settings :: Lens' Job JobSettings

JobSettings contains all the transcode settings for a job.