Network.AWS.MediaConvert.Types.JobTemplate

Description

Documentation

data JobTemplate

A job template is a pre-made set of encoding instructions that you can use to quickly create a job.

See: newJobTemplate smart constructor.

Constructors

JobTemplate'

Fields

accelerationSettings :: Maybe AccelerationSettings

Accelerated transcoding can significantly speed up jobs with long, visually complex content.

lastUpdated :: Maybe POSIX

The timestamp in epoch seconds when the Job template was last updated.

priority :: Maybe Int

Relative priority on the job.

statusUpdateInterval :: Maybe StatusUpdateInterval

Specify how often MediaConvert sends STATUS_UPDATE events to Amazon CloudWatch Events. Set the interval, in seconds, between status updates. MediaConvert sends an update at this interval from the time the service begins processing your job to the time it completes the transcode or encounters an error.

arn :: Maybe Text

An identifier for this resource that is unique within all of AWS.

createdAt :: Maybe POSIX

The timestamp in epoch seconds for Job template creation.

category :: Maybe Text

An optional category you create to organize your job templates.

hopDestinations :: Maybe [HopDestination]

Optional list of hop destinations.

queue :: Maybe Text

Optional. The queue that jobs created from this template are assigned to. If you don't specify this, jobs will go to the default queue.

type' :: Maybe Type

A job template can be of two types: system or custom. System or built-in job templates can't be modified or deleted by the user.

description :: Maybe Text

An optional description you create for each job template.

settings :: JobTemplateSettings

JobTemplateSettings contains all the transcode settings saved in the template that will be applied to jobs created from it.

name :: Text

A name you create for each job template. Each name must be unique within your account.

newJobTemplate

Arguments

:: JobTemplateSettings
$sel:settings:JobTemplate'

-> Text
$sel:name:JobTemplate'

-> JobTemplate

Create a value of JobTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accelerationSettings:JobTemplate', jobTemplate_accelerationSettings - Accelerated transcoding can significantly speed up jobs with long, visually complex content.

$sel:lastUpdated:JobTemplate', jobTemplate_lastUpdated - The timestamp in epoch seconds when the Job template was last updated.

$sel:priority:JobTemplate', jobTemplate_priority - Relative priority on the job.

$sel:statusUpdateInterval:JobTemplate', jobTemplate_statusUpdateInterval - Specify how often MediaConvert sends STATUS_UPDATE events to Amazon CloudWatch Events. Set the interval, in seconds, between status updates. MediaConvert sends an update at this interval from the time the service begins processing your job to the time it completes the transcode or encounters an error.

$sel:arn:JobTemplate', jobTemplate_arn - An identifier for this resource that is unique within all of AWS.

$sel:createdAt:JobTemplate', jobTemplate_createdAt - The timestamp in epoch seconds for Job template creation.

$sel:category:JobTemplate', jobTemplate_category - An optional category you create to organize your job templates.

$sel:hopDestinations:JobTemplate', jobTemplate_hopDestinations - Optional list of hop destinations.

$sel:queue:JobTemplate', jobTemplate_queue - Optional. The queue that jobs created from this template are assigned to. If you don't specify this, jobs will go to the default queue.

$sel:type':JobTemplate', jobTemplate_type - A job template can be of two types: system or custom. System or built-in job templates can't be modified or deleted by the user.

$sel:description:JobTemplate', jobTemplate_description - An optional description you create for each job template.

$sel:settings:JobTemplate', jobTemplate_settings - JobTemplateSettings contains all the transcode settings saved in the template that will be applied to jobs created from it.

$sel:name:JobTemplate', jobTemplate_name - A name you create for each job template. Each name must be unique within your account.

jobTemplate_accelerationSettings :: Lens' JobTemplate (Maybe AccelerationSettings)

Accelerated transcoding can significantly speed up jobs with long, visually complex content.

jobTemplate_lastUpdated :: Lens' JobTemplate (Maybe UTCTime)

The timestamp in epoch seconds when the Job template was last updated.

jobTemplate_priority :: Lens' JobTemplate (Maybe Int)

Relative priority on the job.

jobTemplate_statusUpdateInterval :: Lens' JobTemplate (Maybe StatusUpdateInterval)

Specify how often MediaConvert sends STATUS_UPDATE events to Amazon CloudWatch Events. Set the interval, in seconds, between status updates. MediaConvert sends an update at this interval from the time the service begins processing your job to the time it completes the transcode or encounters an error.

jobTemplate_arn :: Lens' JobTemplate (Maybe Text)

An identifier for this resource that is unique within all of AWS.

jobTemplate_createdAt :: Lens' JobTemplate (Maybe UTCTime)

The timestamp in epoch seconds for Job template creation.

jobTemplate_category :: Lens' JobTemplate (Maybe Text)

An optional category you create to organize your job templates.

jobTemplate_hopDestinations :: Lens' JobTemplate (Maybe [HopDestination])

Optional list of hop destinations.

jobTemplate_queue :: Lens' JobTemplate (Maybe Text)

Optional. The queue that jobs created from this template are assigned to. If you don't specify this, jobs will go to the default queue.

jobTemplate_type :: Lens' JobTemplate (Maybe Type)

A job template can be of two types: system or custom. System or built-in job templates can't be modified or deleted by the user.

jobTemplate_description :: Lens' JobTemplate (Maybe Text)

An optional description you create for each job template.

jobTemplate_settings :: Lens' JobTemplate JobTemplateSettings

JobTemplateSettings contains all the transcode settings saved in the template that will be applied to jobs created from it.

jobTemplate_name :: Lens' JobTemplate Text

A name you create for each job template. Each name must be unique within your account.