Network.AWS.MediaConvert.Types.JobTemplateSettings

Description

Documentation

data JobTemplateSettings

JobTemplateSettings contains all the transcode settings saved in the template that will be applied to jobs created from it.

See: newJobTemplateSettings smart constructor.

Constructors

JobTemplateSettings'

Fields

nielsenNonLinearWatermark :: Maybe NielsenNonLinearWatermarkSettings

Ignore these settings unless you are using Nielsen non-linear watermarking. Specify the values that MediaConvert uses to generate and place Nielsen watermarks in your output audio. In addition to specifying these values, you also need to set up your cloud TIC server. These settings apply to every output in your job. The MediaConvert implementation is currently with the following Nielsen versions: Nielsen Watermark SDK Version 5.2.1 Nielsen NLM Watermark Engine Version 1.2.7 Nielsen Watermark Authenticator [SID_TIC] Version [5.0.0]

esam :: Maybe EsamSettings

Settings for Event Signaling And Messaging (ESAM). If you don't do ad insertion, you can ignore these settings.

inputs :: Maybe [InputTemplate]

Use Inputs (inputs) to define the source file used in the transcode job. There can only be one input in a job template. Using the API, you can include multiple inputs when referencing a job template.

timedMetadataInsertion :: Maybe TimedMetadataInsertion

Enable Timed metadata insertion (TimedMetadataInsertion) to include ID3 tags in any HLS outputs. To include timed metadata, you must enable it here, enable it in each output container, and specify tags and timecodes in ID3 insertion (Id3Insertion) objects.

nielsenConfiguration :: Maybe NielsenConfiguration

Settings for your Nielsen configuration. If you don't do Nielsen measurement and analytics, ignore these settings. When you enable Nielsen configuration (nielsenConfiguration), MediaConvert enables PCM to ID3 tagging for all outputs in the job. To enable Nielsen configuration programmatically, include an instance of nielsenConfiguration in your JSON job specification. Even if you don't include any children of nielsenConfiguration, you still enable the setting.

availBlanking :: Maybe AvailBlanking

Settings for ad avail blanking. Video can be blanked or overlaid with an image, and audio muted during SCTE-35 triggered ad avails.

extendedDataServices :: Maybe ExtendedDataServices

If your source content has EIA-608 Line 21 Data Services, enable this feature to specify what MediaConvert does with the Extended Data Services (XDS) packets. You can choose to pass through XDS packets, or remove them from the output. For more information about XDS, see EIA-608 Line Data Services, section 9.5.1.5 05h Content Advisory.

motionImageInserter :: Maybe MotionImageInserter

Overlay motion graphics on top of your video. The motion graphics that you specify here appear on all outputs in all output groups. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/motion-graphic-overlay.html.

timecodeConfig :: Maybe TimecodeConfig

These settings control how the service handles timecodes throughout the job. These settings don't affect input clipping.

outputGroups :: Maybe [OutputGroup]

(OutputGroups) contains one group of settings for each set of outputs that share a common package type. All unpackaged files (MPEG-4, MPEG-2 TS, Quicktime, MXF, and no container) are grouped in a single output group as well. Required in (OutputGroups) is a group of settings that apply to the whole group. This required object depends on the value you set for (Type) under (OutputGroups)>(OutputGroupSettings). Type, settings object pairs are as follows. * FILE_GROUP_SETTINGS, FileGroupSettings * HLS_GROUP_SETTINGS, HlsGroupSettings * DASH_ISO_GROUP_SETTINGS, DashIsoGroupSettings * MS_SMOOTH_GROUP_SETTINGS, MsSmoothGroupSettings * CMAF_GROUP_SETTINGS, CmafGroupSettings

adAvailOffset :: Maybe Int

When specified, this offset (in milliseconds) is added to the input Ad Avail PTS time.

kantarWatermark :: Maybe KantarWatermarkSettings

Use these settings only when you use Kantar watermarking. Specify the values that MediaConvert uses to generate and place Kantar watermarks in your output audio. These settings apply to every output in your job. In addition to specifying these values, you also need to store your Kantar credentials in AWS Secrets Manager. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/kantar-watermarking.html.

newJobTemplateSettings :: JobTemplateSettings

Create a value of JobTemplateSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nielsenNonLinearWatermark:JobTemplateSettings', jobTemplateSettings_nielsenNonLinearWatermark - Ignore these settings unless you are using Nielsen non-linear watermarking. Specify the values that MediaConvert uses to generate and place Nielsen watermarks in your output audio. In addition to specifying these values, you also need to set up your cloud TIC server. These settings apply to every output in your job. The MediaConvert implementation is currently with the following Nielsen versions: Nielsen Watermark SDK Version 5.2.1 Nielsen NLM Watermark Engine Version 1.2.7 Nielsen Watermark Authenticator [SID_TIC] Version [5.0.0]

$sel:esam:JobTemplateSettings', jobTemplateSettings_esam - Settings for Event Signaling And Messaging (ESAM). If you don't do ad insertion, you can ignore these settings.

$sel:inputs:JobTemplateSettings', jobTemplateSettings_inputs - Use Inputs (inputs) to define the source file used in the transcode job. There can only be one input in a job template. Using the API, you can include multiple inputs when referencing a job template.

$sel:timedMetadataInsertion:JobTemplateSettings', jobTemplateSettings_timedMetadataInsertion - Enable Timed metadata insertion (TimedMetadataInsertion) to include ID3 tags in any HLS outputs. To include timed metadata, you must enable it here, enable it in each output container, and specify tags and timecodes in ID3 insertion (Id3Insertion) objects.

$sel:nielsenConfiguration:JobTemplateSettings', jobTemplateSettings_nielsenConfiguration - Settings for your Nielsen configuration. If you don't do Nielsen measurement and analytics, ignore these settings. When you enable Nielsen configuration (nielsenConfiguration), MediaConvert enables PCM to ID3 tagging for all outputs in the job. To enable Nielsen configuration programmatically, include an instance of nielsenConfiguration in your JSON job specification. Even if you don't include any children of nielsenConfiguration, you still enable the setting.

$sel:availBlanking:JobTemplateSettings', jobTemplateSettings_availBlanking - Settings for ad avail blanking. Video can be blanked or overlaid with an image, and audio muted during SCTE-35 triggered ad avails.

$sel:extendedDataServices:JobTemplateSettings', jobTemplateSettings_extendedDataServices - If your source content has EIA-608 Line 21 Data Services, enable this feature to specify what MediaConvert does with the Extended Data Services (XDS) packets. You can choose to pass through XDS packets, or remove them from the output. For more information about XDS, see EIA-608 Line Data Services, section 9.5.1.5 05h Content Advisory.

$sel:motionImageInserter:JobTemplateSettings', jobTemplateSettings_motionImageInserter - Overlay motion graphics on top of your video. The motion graphics that you specify here appear on all outputs in all output groups. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/motion-graphic-overlay.html.

$sel:timecodeConfig:JobTemplateSettings', jobTemplateSettings_timecodeConfig - These settings control how the service handles timecodes throughout the job. These settings don't affect input clipping.

$sel:outputGroups:JobTemplateSettings', jobTemplateSettings_outputGroups - (OutputGroups) contains one group of settings for each set of outputs that share a common package type. All unpackaged files (MPEG-4, MPEG-2 TS, Quicktime, MXF, and no container) are grouped in a single output group as well. Required in (OutputGroups) is a group of settings that apply to the whole group. This required object depends on the value you set for (Type) under (OutputGroups)>(OutputGroupSettings). Type, settings object pairs are as follows. * FILE_GROUP_SETTINGS, FileGroupSettings * HLS_GROUP_SETTINGS, HlsGroupSettings * DASH_ISO_GROUP_SETTINGS, DashIsoGroupSettings * MS_SMOOTH_GROUP_SETTINGS, MsSmoothGroupSettings * CMAF_GROUP_SETTINGS, CmafGroupSettings

$sel:adAvailOffset:JobTemplateSettings', jobTemplateSettings_adAvailOffset - When specified, this offset (in milliseconds) is added to the input Ad Avail PTS time.

$sel:kantarWatermark:JobTemplateSettings', jobTemplateSettings_kantarWatermark - Use these settings only when you use Kantar watermarking. Specify the values that MediaConvert uses to generate and place Kantar watermarks in your output audio. These settings apply to every output in your job. In addition to specifying these values, you also need to store your Kantar credentials in AWS Secrets Manager. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/kantar-watermarking.html.

jobTemplateSettings_nielsenNonLinearWatermark :: Lens' JobTemplateSettings (Maybe NielsenNonLinearWatermarkSettings)

Ignore these settings unless you are using Nielsen non-linear watermarking. Specify the values that MediaConvert uses to generate and place Nielsen watermarks in your output audio. In addition to specifying these values, you also need to set up your cloud TIC server. These settings apply to every output in your job. The MediaConvert implementation is currently with the following Nielsen versions: Nielsen Watermark SDK Version 5.2.1 Nielsen NLM Watermark Engine Version 1.2.7 Nielsen Watermark Authenticator [SID_TIC] Version [5.0.0]

jobTemplateSettings_esam :: Lens' JobTemplateSettings (Maybe EsamSettings)

Settings for Event Signaling And Messaging (ESAM). If you don't do ad insertion, you can ignore these settings.

jobTemplateSettings_inputs :: Lens' JobTemplateSettings (Maybe [InputTemplate])

Use Inputs (inputs) to define the source file used in the transcode job. There can only be one input in a job template. Using the API, you can include multiple inputs when referencing a job template.

jobTemplateSettings_timedMetadataInsertion :: Lens' JobTemplateSettings (Maybe TimedMetadataInsertion)

Enable Timed metadata insertion (TimedMetadataInsertion) to include ID3 tags in any HLS outputs. To include timed metadata, you must enable it here, enable it in each output container, and specify tags and timecodes in ID3 insertion (Id3Insertion) objects.

jobTemplateSettings_nielsenConfiguration :: Lens' JobTemplateSettings (Maybe NielsenConfiguration)

Settings for your Nielsen configuration. If you don't do Nielsen measurement and analytics, ignore these settings. When you enable Nielsen configuration (nielsenConfiguration), MediaConvert enables PCM to ID3 tagging for all outputs in the job. To enable Nielsen configuration programmatically, include an instance of nielsenConfiguration in your JSON job specification. Even if you don't include any children of nielsenConfiguration, you still enable the setting.

jobTemplateSettings_availBlanking :: Lens' JobTemplateSettings (Maybe AvailBlanking)

Settings for ad avail blanking. Video can be blanked or overlaid with an image, and audio muted during SCTE-35 triggered ad avails.

jobTemplateSettings_extendedDataServices :: Lens' JobTemplateSettings (Maybe ExtendedDataServices)

If your source content has EIA-608 Line 21 Data Services, enable this feature to specify what MediaConvert does with the Extended Data Services (XDS) packets. You can choose to pass through XDS packets, or remove them from the output. For more information about XDS, see EIA-608 Line Data Services, section 9.5.1.5 05h Content Advisory.

jobTemplateSettings_motionImageInserter :: Lens' JobTemplateSettings (Maybe MotionImageInserter)

Overlay motion graphics on top of your video. The motion graphics that you specify here appear on all outputs in all output groups. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/motion-graphic-overlay.html.

jobTemplateSettings_timecodeConfig :: Lens' JobTemplateSettings (Maybe TimecodeConfig)

These settings control how the service handles timecodes throughout the job. These settings don't affect input clipping.

jobTemplateSettings_outputGroups :: Lens' JobTemplateSettings (Maybe [OutputGroup])

(OutputGroups) contains one group of settings for each set of outputs that share a common package type. All unpackaged files (MPEG-4, MPEG-2 TS, Quicktime, MXF, and no container) are grouped in a single output group as well. Required in (OutputGroups) is a group of settings that apply to the whole group. This required object depends on the value you set for (Type) under (OutputGroups)>(OutputGroupSettings). Type, settings object pairs are as follows. * FILE_GROUP_SETTINGS, FileGroupSettings * HLS_GROUP_SETTINGS, HlsGroupSettings * DASH_ISO_GROUP_SETTINGS, DashIsoGroupSettings * MS_SMOOTH_GROUP_SETTINGS, MsSmoothGroupSettings * CMAF_GROUP_SETTINGS, CmafGroupSettings

jobTemplateSettings_adAvailOffset :: Lens' JobTemplateSettings (Maybe Int)

When specified, this offset (in milliseconds) is added to the input Ad Avail PTS time.

jobTemplateSettings_kantarWatermark :: Lens' JobTemplateSettings (Maybe KantarWatermarkSettings)

Use these settings only when you use Kantar watermarking. Specify the values that MediaConvert uses to generate and place Kantar watermarks in your output audio. These settings apply to every output in your job. In addition to specifying these values, you also need to store your Kantar credentials in AWS Secrets Manager. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/kantar-watermarking.html.