Network.AWS.MediaConvert.Types.KantarWatermarkSettings

Description

Documentation

data KantarWatermarkSettings

Use these settings only when you use Kantar watermarking. Specify the values that MediaConvert uses to generate and place Kantar watermarks in your output audio. These settings apply to every output in your job. In addition to specifying these values, you also need to store your Kantar credentials in AWS Secrets Manager. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/kantar-watermarking.html.

See: newKantarWatermarkSettings smart constructor.

Constructors

KantarWatermarkSettings'

Fields

channelName :: Maybe Text

Provide an audio channel name from your Kantar audio license.

metadata4 :: Maybe Text

Additional metadata that MediaConvert sends to Kantar. Maximum length is 50 characters.

metadata3 :: Maybe Text

You can optionally use this field to specify the first timestamp that Kantar embeds during watermarking. Kantar suggests that you be very cautious when using this Kantar feature, and that you use it only on channels that are managed specifically for use with this feature by your Audience Measurement Operator. For more information about this feature, contact Kantar technical support.

fileOffset :: Maybe Double

Optional. Specify an offset, in whole seconds, from the start of your output and the beginning of the watermarking. When you don't specify an offset, Kantar defaults to zero.

metadata5 :: Maybe Text

Additional metadata that MediaConvert sends to Kantar. Maximum length is 50 characters.

logDestination :: Maybe Text

Optional. Specify the Amazon S3 bucket where you want MediaConvert to store your Kantar watermark XML logs. When you don't specify a bucket, MediaConvert doesn't save these logs. Note that your MediaConvert service role must provide access to this location. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/iam-role.html

kantarServerUrl :: Maybe Text

Provide the HTTPS endpoint to the Kantar server. You should get this endpoint from Kantar.

kantarLicenseId :: Maybe Natural

Provide your Kantar license ID number. You should get this number from Kantar.

credentialsSecretName :: Maybe Text

Provide the name of the AWS Secrets Manager secret where your Kantar credentials are stored. Note that your MediaConvert service role must provide access to this secret. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/granting-permissions-for-mediaconvert-to-access-secrets-manager-secret.html. For instructions on creating a secret, see https://docs.aws.amazon.com/secretsmanager/latest/userguide/tutorials_basic.html, in the AWS Secrets Manager User Guide.

metadata6 :: Maybe Text

Additional metadata that MediaConvert sends to Kantar. Maximum length is 50 characters.

metadata8 :: Maybe Text

Additional metadata that MediaConvert sends to Kantar. Maximum length is 50 characters.

metadata7 :: Maybe Text

Additional metadata that MediaConvert sends to Kantar. Maximum length is 50 characters.

contentReference :: Maybe Text

Specify a unique identifier for Kantar to use for this piece of content.

newKantarWatermarkSettings :: KantarWatermarkSettings

Create a value of KantarWatermarkSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelName:KantarWatermarkSettings', kantarWatermarkSettings_channelName - Provide an audio channel name from your Kantar audio license.

$sel:metadata4:KantarWatermarkSettings', kantarWatermarkSettings_metadata4 - Additional metadata that MediaConvert sends to Kantar. Maximum length is 50 characters.

$sel:metadata3:KantarWatermarkSettings', kantarWatermarkSettings_metadata3 - You can optionally use this field to specify the first timestamp that Kantar embeds during watermarking. Kantar suggests that you be very cautious when using this Kantar feature, and that you use it only on channels that are managed specifically for use with this feature by your Audience Measurement Operator. For more information about this feature, contact Kantar technical support.

$sel:fileOffset:KantarWatermarkSettings', kantarWatermarkSettings_fileOffset - Optional. Specify an offset, in whole seconds, from the start of your output and the beginning of the watermarking. When you don't specify an offset, Kantar defaults to zero.

$sel:metadata5:KantarWatermarkSettings', kantarWatermarkSettings_metadata5 - Additional metadata that MediaConvert sends to Kantar. Maximum length is 50 characters.

$sel:logDestination:KantarWatermarkSettings', kantarWatermarkSettings_logDestination - Optional. Specify the Amazon S3 bucket where you want MediaConvert to store your Kantar watermark XML logs. When you don't specify a bucket, MediaConvert doesn't save these logs. Note that your MediaConvert service role must provide access to this location. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/iam-role.html

$sel:kantarServerUrl:KantarWatermarkSettings', kantarWatermarkSettings_kantarServerUrl - Provide the HTTPS endpoint to the Kantar server. You should get this endpoint from Kantar.

$sel:kantarLicenseId:KantarWatermarkSettings', kantarWatermarkSettings_kantarLicenseId - Provide your Kantar license ID number. You should get this number from Kantar.

$sel:credentialsSecretName:KantarWatermarkSettings', kantarWatermarkSettings_credentialsSecretName - Provide the name of the AWS Secrets Manager secret where your Kantar credentials are stored. Note that your MediaConvert service role must provide access to this secret. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/granting-permissions-for-mediaconvert-to-access-secrets-manager-secret.html. For instructions on creating a secret, see https://docs.aws.amazon.com/secretsmanager/latest/userguide/tutorials_basic.html, in the AWS Secrets Manager User Guide.

$sel:metadata6:KantarWatermarkSettings', kantarWatermarkSettings_metadata6 - Additional metadata that MediaConvert sends to Kantar. Maximum length is 50 characters.

$sel:metadata8:KantarWatermarkSettings', kantarWatermarkSettings_metadata8 - Additional metadata that MediaConvert sends to Kantar. Maximum length is 50 characters.

$sel:metadata7:KantarWatermarkSettings', kantarWatermarkSettings_metadata7 - Additional metadata that MediaConvert sends to Kantar. Maximum length is 50 characters.

$sel:contentReference:KantarWatermarkSettings', kantarWatermarkSettings_contentReference - Specify a unique identifier for Kantar to use for this piece of content.

kantarWatermarkSettings_channelName :: Lens' KantarWatermarkSettings (Maybe Text)

Provide an audio channel name from your Kantar audio license.

kantarWatermarkSettings_metadata4 :: Lens' KantarWatermarkSettings (Maybe Text)

Additional metadata that MediaConvert sends to Kantar. Maximum length is 50 characters.

kantarWatermarkSettings_metadata3 :: Lens' KantarWatermarkSettings (Maybe Text)

You can optionally use this field to specify the first timestamp that Kantar embeds during watermarking. Kantar suggests that you be very cautious when using this Kantar feature, and that you use it only on channels that are managed specifically for use with this feature by your Audience Measurement Operator. For more information about this feature, contact Kantar technical support.

kantarWatermarkSettings_fileOffset :: Lens' KantarWatermarkSettings (Maybe Double)

Optional. Specify an offset, in whole seconds, from the start of your output and the beginning of the watermarking. When you don't specify an offset, Kantar defaults to zero.

kantarWatermarkSettings_metadata5 :: Lens' KantarWatermarkSettings (Maybe Text)

Additional metadata that MediaConvert sends to Kantar. Maximum length is 50 characters.

kantarWatermarkSettings_logDestination :: Lens' KantarWatermarkSettings (Maybe Text)

Optional. Specify the Amazon S3 bucket where you want MediaConvert to store your Kantar watermark XML logs. When you don't specify a bucket, MediaConvert doesn't save these logs. Note that your MediaConvert service role must provide access to this location. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/iam-role.html

kantarWatermarkSettings_kantarServerUrl :: Lens' KantarWatermarkSettings (Maybe Text)

Provide the HTTPS endpoint to the Kantar server. You should get this endpoint from Kantar.

kantarWatermarkSettings_kantarLicenseId :: Lens' KantarWatermarkSettings (Maybe Natural)

Provide your Kantar license ID number. You should get this number from Kantar.

kantarWatermarkSettings_credentialsSecretName :: Lens' KantarWatermarkSettings (Maybe Text)

Provide the name of the AWS Secrets Manager secret where your Kantar credentials are stored. Note that your MediaConvert service role must provide access to this secret. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/granting-permissions-for-mediaconvert-to-access-secrets-manager-secret.html. For instructions on creating a secret, see https://docs.aws.amazon.com/secretsmanager/latest/userguide/tutorials_basic.html, in the AWS Secrets Manager User Guide.

kantarWatermarkSettings_metadata6 :: Lens' KantarWatermarkSettings (Maybe Text)

Additional metadata that MediaConvert sends to Kantar. Maximum length is 50 characters.

kantarWatermarkSettings_metadata8 :: Lens' KantarWatermarkSettings (Maybe Text)

Additional metadata that MediaConvert sends to Kantar. Maximum length is 50 characters.

kantarWatermarkSettings_metadata7 :: Lens' KantarWatermarkSettings (Maybe Text)

Additional metadata that MediaConvert sends to Kantar. Maximum length is 50 characters.

kantarWatermarkSettings_contentReference :: Lens' KantarWatermarkSettings (Maybe Text)

Specify a unique identifier for Kantar to use for this piece of content.