Network.AWS.MediaConvert.Types.MovSettings

Description

Documentation

data MovSettings

These settings relate to your QuickTime MOV output container.

See: newMovSettings smart constructor.

Constructors

MovSettings'

Fields

paddingControl :: Maybe MovPaddingControl

To make this output compatible with Omenon, keep the default value, OMNEON. Unless you need Omneon compatibility, set this value to NONE. When you keep the default value, OMNEON, MediaConvert increases the length of the edit list atom. This might cause file rejections when a recipient of the output file doesn't expct this extra padding.

mpeg2FourCCControl :: Maybe MovMpeg2FourCCControl

When set to XDCAM, writes MPEG2 video streams into the QuickTime file using XDCAM fourcc codes. This increases compatibility with Apple editors and players, but may decrease compatibility with other players. Only applicable when the video codec is MPEG2.

cslgAtom :: Maybe MovCslgAtom

When enabled, file composition times will start at zero, composition times in the 'ctts' (composition time to sample) box for B-frames will be negative, and a 'cslg' (composition shift least greatest) box will be included per 14496-1 amendment 1. This improves compatibility with Apple players and tools.

clapAtom :: Maybe MovClapAtom

When enabled, include 'clap' atom if appropriate for the video output settings.

reference :: Maybe MovReference

Always keep the default value (SELF_CONTAINED) for this setting.

newMovSettings :: MovSettings

Create a value of MovSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:paddingControl:MovSettings', movSettings_paddingControl - To make this output compatible with Omenon, keep the default value, OMNEON. Unless you need Omneon compatibility, set this value to NONE. When you keep the default value, OMNEON, MediaConvert increases the length of the edit list atom. This might cause file rejections when a recipient of the output file doesn't expct this extra padding.

$sel:mpeg2FourCCControl:MovSettings', movSettings_mpeg2FourCCControl - When set to XDCAM, writes MPEG2 video streams into the QuickTime file using XDCAM fourcc codes. This increases compatibility with Apple editors and players, but may decrease compatibility with other players. Only applicable when the video codec is MPEG2.

$sel:cslgAtom:MovSettings', movSettings_cslgAtom - When enabled, file composition times will start at zero, composition times in the 'ctts' (composition time to sample) box for B-frames will be negative, and a 'cslg' (composition shift least greatest) box will be included per 14496-1 amendment 1. This improves compatibility with Apple players and tools.

$sel:clapAtom:MovSettings', movSettings_clapAtom - When enabled, include 'clap' atom if appropriate for the video output settings.

$sel:reference:MovSettings', movSettings_reference - Always keep the default value (SELF_CONTAINED) for this setting.

movSettings_paddingControl :: Lens' MovSettings (Maybe MovPaddingControl)

To make this output compatible with Omenon, keep the default value, OMNEON. Unless you need Omneon compatibility, set this value to NONE. When you keep the default value, OMNEON, MediaConvert increases the length of the edit list atom. This might cause file rejections when a recipient of the output file doesn't expct this extra padding.

movSettings_mpeg2FourCCControl :: Lens' MovSettings (Maybe MovMpeg2FourCCControl)

When set to XDCAM, writes MPEG2 video streams into the QuickTime file using XDCAM fourcc codes. This increases compatibility with Apple editors and players, but may decrease compatibility with other players. Only applicable when the video codec is MPEG2.

movSettings_cslgAtom :: Lens' MovSettings (Maybe MovCslgAtom)

When enabled, file composition times will start at zero, composition times in the 'ctts' (composition time to sample) box for B-frames will be negative, and a 'cslg' (composition shift least greatest) box will be included per 14496-1 amendment 1. This improves compatibility with Apple players and tools.

movSettings_clapAtom :: Lens' MovSettings (Maybe MovClapAtom)

When enabled, include 'clap' atom if appropriate for the video output settings.

movSettings_reference :: Lens' MovSettings (Maybe MovReference)

Always keep the default value (SELF_CONTAINED) for this setting.