Network.AWS.MediaConvert.Types.MsSmoothGroupSettings

Description

Documentation

data MsSmoothGroupSettings

Settings related to your Microsoft Smooth Streaming output package. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/outputs-file-ABR.html. When you work directly in your JSON job specification, include this object and any required children when you set Type, under OutputGroupSettings, to MS_SMOOTH_GROUP_SETTINGS.

See: newMsSmoothGroupSettings smart constructor.

Constructors

MsSmoothGroupSettings'

Fields

manifestEncoding :: Maybe MsSmoothManifestEncoding

Use Manifest encoding (MsSmoothManifestEncoding) to specify the encoding format for the server and client manifest. Valid options are utf8 and utf16.

fragmentLength :: Maybe Natural

Specify how you want MediaConvert to determine the fragment length. Choose Exact (EXACT) to have the encoder use the exact length that you specify with the setting Fragment length (FragmentLength). This might result in extra I-frames. Choose Multiple of GOP (GOP_MULTIPLE) to have the encoder round up the segment lengths to match the next GOP boundary.

additionalManifests :: Maybe [MsSmoothAdditionalManifest]

By default, the service creates one .ism Microsoft Smooth Streaming manifest for each Microsoft Smooth Streaming output group in your job. This default manifest references every output in the output group. To create additional manifests that reference a subset of the outputs in the output group, specify a list of them here.

encryption :: Maybe MsSmoothEncryptionSettings

If you are using DRM, set DRM System (MsSmoothEncryptionSettings) to specify the value SpekeKeyProvider.

destination :: Maybe Text

Use Destination (Destination) to specify the S3 output location and the output filename base. Destination accepts format identifiers. If you do not specify the base filename in the URI, the service will use the filename of the input file. If your job has multiple inputs, the service uses the filename of the first input file.

destinationSettings :: Maybe DestinationSettings

Settings associated with the destination. Will vary based on the type of destination

fragmentLengthControl :: Maybe MsSmoothFragmentLengthControl

Specify how you want MediaConvert to determine the fragment length. Choose Exact (EXACT) to have the encoder use the exact length that you specify with the setting Fragment length (FragmentLength). This might result in extra I-frames. Choose Multiple of GOP (GOP_MULTIPLE) to have the encoder round up the segment lengths to match the next GOP boundary.

audioDeduplication :: Maybe MsSmoothAudioDeduplication

COMBINE_DUPLICATE_STREAMS combines identical audio encoding settings across a Microsoft Smooth output group into a single audio stream.

newMsSmoothGroupSettings :: MsSmoothGroupSettings

Create a value of MsSmoothGroupSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:manifestEncoding:MsSmoothGroupSettings', msSmoothGroupSettings_manifestEncoding - Use Manifest encoding (MsSmoothManifestEncoding) to specify the encoding format for the server and client manifest. Valid options are utf8 and utf16.

$sel:fragmentLength:MsSmoothGroupSettings', msSmoothGroupSettings_fragmentLength - Specify how you want MediaConvert to determine the fragment length. Choose Exact (EXACT) to have the encoder use the exact length that you specify with the setting Fragment length (FragmentLength). This might result in extra I-frames. Choose Multiple of GOP (GOP_MULTIPLE) to have the encoder round up the segment lengths to match the next GOP boundary.

$sel:additionalManifests:MsSmoothGroupSettings', msSmoothGroupSettings_additionalManifests - By default, the service creates one .ism Microsoft Smooth Streaming manifest for each Microsoft Smooth Streaming output group in your job. This default manifest references every output in the output group. To create additional manifests that reference a subset of the outputs in the output group, specify a list of them here.

$sel:encryption:MsSmoothGroupSettings', msSmoothGroupSettings_encryption - If you are using DRM, set DRM System (MsSmoothEncryptionSettings) to specify the value SpekeKeyProvider.

$sel:destination:MsSmoothGroupSettings', msSmoothGroupSettings_destination - Use Destination (Destination) to specify the S3 output location and the output filename base. Destination accepts format identifiers. If you do not specify the base filename in the URI, the service will use the filename of the input file. If your job has multiple inputs, the service uses the filename of the first input file.

$sel:destinationSettings:MsSmoothGroupSettings', msSmoothGroupSettings_destinationSettings - Settings associated with the destination. Will vary based on the type of destination

$sel:fragmentLengthControl:MsSmoothGroupSettings', msSmoothGroupSettings_fragmentLengthControl - Specify how you want MediaConvert to determine the fragment length. Choose Exact (EXACT) to have the encoder use the exact length that you specify with the setting Fragment length (FragmentLength). This might result in extra I-frames. Choose Multiple of GOP (GOP_MULTIPLE) to have the encoder round up the segment lengths to match the next GOP boundary.

$sel:audioDeduplication:MsSmoothGroupSettings', msSmoothGroupSettings_audioDeduplication - COMBINE_DUPLICATE_STREAMS combines identical audio encoding settings across a Microsoft Smooth output group into a single audio stream.

msSmoothGroupSettings_manifestEncoding :: Lens' MsSmoothGroupSettings (Maybe MsSmoothManifestEncoding)

Use Manifest encoding (MsSmoothManifestEncoding) to specify the encoding format for the server and client manifest. Valid options are utf8 and utf16.

msSmoothGroupSettings_fragmentLength :: Lens' MsSmoothGroupSettings (Maybe Natural)

Specify how you want MediaConvert to determine the fragment length. Choose Exact (EXACT) to have the encoder use the exact length that you specify with the setting Fragment length (FragmentLength). This might result in extra I-frames. Choose Multiple of GOP (GOP_MULTIPLE) to have the encoder round up the segment lengths to match the next GOP boundary.

msSmoothGroupSettings_additionalManifests :: Lens' MsSmoothGroupSettings (Maybe [MsSmoothAdditionalManifest])

By default, the service creates one .ism Microsoft Smooth Streaming manifest for each Microsoft Smooth Streaming output group in your job. This default manifest references every output in the output group. To create additional manifests that reference a subset of the outputs in the output group, specify a list of them here.

msSmoothGroupSettings_encryption :: Lens' MsSmoothGroupSettings (Maybe MsSmoothEncryptionSettings)

If you are using DRM, set DRM System (MsSmoothEncryptionSettings) to specify the value SpekeKeyProvider.

msSmoothGroupSettings_destination :: Lens' MsSmoothGroupSettings (Maybe Text)

Use Destination (Destination) to specify the S3 output location and the output filename base. Destination accepts format identifiers. If you do not specify the base filename in the URI, the service will use the filename of the input file. If your job has multiple inputs, the service uses the filename of the first input file.

msSmoothGroupSettings_destinationSettings :: Lens' MsSmoothGroupSettings (Maybe DestinationSettings)

Settings associated with the destination. Will vary based on the type of destination

msSmoothGroupSettings_fragmentLengthControl :: Lens' MsSmoothGroupSettings (Maybe MsSmoothFragmentLengthControl)

Specify how you want MediaConvert to determine the fragment length. Choose Exact (EXACT) to have the encoder use the exact length that you specify with the setting Fragment length (FragmentLength). This might result in extra I-frames. Choose Multiple of GOP (GOP_MULTIPLE) to have the encoder round up the segment lengths to match the next GOP boundary.

msSmoothGroupSettings_audioDeduplication :: Lens' MsSmoothGroupSettings (Maybe MsSmoothAudioDeduplication)

COMBINE_DUPLICATE_STREAMS combines identical audio encoding settings across a Microsoft Smooth output group into a single audio stream.