Network.AWS.MediaConvert.Types.NielsenConfiguration

Description

Documentation

data NielsenConfiguration

Settings for your Nielsen configuration. If you don't do Nielsen measurement and analytics, ignore these settings. When you enable Nielsen configuration (nielsenConfiguration), MediaConvert enables PCM to ID3 tagging for all outputs in the job. To enable Nielsen configuration programmatically, include an instance of nielsenConfiguration in your JSON job specification. Even if you don't include any children of nielsenConfiguration, you still enable the setting.

See: newNielsenConfiguration smart constructor.

Constructors

NielsenConfiguration'

Fields

breakoutCode :: Maybe Natural

Nielsen has discontinued the use of breakout code functionality. If you must include this property, set the value to zero.

distributorId :: Maybe Text

Use Distributor ID (DistributorID) to specify the distributor ID that is assigned to your organization by Neilsen.

newNielsenConfiguration :: NielsenConfiguration

Create a value of NielsenConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:breakoutCode:NielsenConfiguration', nielsenConfiguration_breakoutCode - Nielsen has discontinued the use of breakout code functionality. If you must include this property, set the value to zero.

$sel:distributorId:NielsenConfiguration', nielsenConfiguration_distributorId - Use Distributor ID (DistributorID) to specify the distributor ID that is assigned to your organization by Neilsen.

nielsenConfiguration_breakoutCode :: Lens' NielsenConfiguration (Maybe Natural)

Nielsen has discontinued the use of breakout code functionality. If you must include this property, set the value to zero.

nielsenConfiguration_distributorId :: Lens' NielsenConfiguration (Maybe Text)

Use Distributor ID (DistributorID) to specify the distributor ID that is assigned to your organization by Neilsen.