Network.AWS.MediaConvert.Types.NielsenNonLinearWatermarkSettings

Description

Documentation

data NielsenNonLinearWatermarkSettings

Ignore these settings unless you are using Nielsen non-linear watermarking. Specify the values that MediaConvert uses to generate and place Nielsen watermarks in your output audio. In addition to specifying these values, you also need to set up your cloud TIC server. These settings apply to every output in your job. The MediaConvert implementation is currently with the following Nielsen versions: Nielsen Watermark SDK Version 5.2.1 Nielsen NLM Watermark Engine Version 1.2.7 Nielsen Watermark Authenticator [SID_TIC] Version [5.0.0]

See: newNielsenNonLinearWatermarkSettings smart constructor.

Constructors

NielsenNonLinearWatermarkSettings'

Fields

episodeId :: Maybe Text

Optional. If this asset uses an episode ID with Nielsen, provide it here.

activeWatermarkProcess :: Maybe NielsenActiveWatermarkProcessType

Choose the type of Nielsen watermarks that you want in your outputs. When you choose NAES 2 and NW (NAES2_AND_NW), you must provide a value for the setting SID (sourceId). When you choose CBET (CBET), you must provide a value for the setting CSID (cbetSourceId). When you choose NAES 2, NW, and CBET (NAES2_AND_NW_AND_CBET), you must provide values for both of these settings.

sourceId :: Maybe Natural

Use the SID that Nielsen provides to you. This source ID should be unique to your Nielsen account but common to all of your output assets. Required for all Nielsen non-linear watermarking. This ID should be unique to your Nielsen account but common to all of your output assets. Required for all Nielsen non-linear watermarking.

cbetSourceId :: Maybe Text

Use the CSID that Nielsen provides to you. This CBET source ID should be unique to your Nielsen account but common to all of your output assets that have CBET watermarking. Required when you choose a value for the setting Watermark types (ActiveWatermarkProcess) that includes CBET.

ticServerUrl :: Maybe Text

Specify the endpoint for the TIC server that you have deployed and configured in the AWS Cloud. Required for all Nielsen non-linear watermarking. MediaConvert can't connect directly to a TIC server. Instead, you must use API Gateway to provide a RESTful interface between MediaConvert and a TIC server that you deploy in your AWS account. For more information on deploying a TIC server in your AWS account and the required API Gateway, contact Nielsen support.

metadataDestination :: Maybe Text

Specify the Amazon S3 location where you want MediaConvert to save your Nielsen non-linear metadata .zip file. This Amazon S3 bucket must be in the same Region as the one where you do your MediaConvert transcoding. If you want to include an ADI file in this .zip file, use the setting ADI file (adiFilename) to specify it. MediaConvert delivers the Nielsen metadata .zip files only to your metadata destination Amazon S3 bucket. It doesn't deliver the .zip files to Nielsen. You are responsible for delivering the metadata .zip files to Nielsen.

assetName :: Maybe Text

Use the asset name that you provide to Nielsen for this asset. Required for all Nielsen non-linear watermarking.

adiFilename :: Maybe Text

Optional. Use this setting when you want the service to include an ADI file in the Nielsen metadata .zip file. To provide an ADI file, store it in Amazon S3 and provide a URL to it here. The URL should be in the following format: S3://bucket/path/ADI-file. For more information about the metadata .zip file, see the setting Metadata destination (metadataDestination).

assetId :: Maybe Text

Use the asset ID that you provide to Nielsen to uniquely identify this asset. Required for all Nielsen non-linear watermarking.

uniqueTicPerAudioTrack :: Maybe NielsenUniqueTicPerAudioTrackType

To create assets that have the same TIC values in each audio track, keep the default value Share TICs (SAME_TICS_PER_TRACK). To create assets that have unique TIC values for each audio track, choose Use unique TICs (RESERVE_UNIQUE_TICS_PER_TRACK).

sourceWatermarkStatus :: Maybe NielsenSourceWatermarkStatusType

Required. Specify whether your source content already contains Nielsen non-linear watermarks. When you set this value to Watermarked (WATERMARKED), the service fails the job. Nielsen requires that you add non-linear watermarking to only clean content that doesn't already have non-linear Nielsen watermarks.

newNielsenNonLinearWatermarkSettings :: NielsenNonLinearWatermarkSettings

Create a value of NielsenNonLinearWatermarkSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:episodeId:NielsenNonLinearWatermarkSettings', nielsenNonLinearWatermarkSettings_episodeId - Optional. If this asset uses an episode ID with Nielsen, provide it here.

$sel:activeWatermarkProcess:NielsenNonLinearWatermarkSettings', nielsenNonLinearWatermarkSettings_activeWatermarkProcess - Choose the type of Nielsen watermarks that you want in your outputs. When you choose NAES 2 and NW (NAES2_AND_NW), you must provide a value for the setting SID (sourceId). When you choose CBET (CBET), you must provide a value for the setting CSID (cbetSourceId). When you choose NAES 2, NW, and CBET (NAES2_AND_NW_AND_CBET), you must provide values for both of these settings.

$sel:sourceId:NielsenNonLinearWatermarkSettings', nielsenNonLinearWatermarkSettings_sourceId - Use the SID that Nielsen provides to you. This source ID should be unique to your Nielsen account but common to all of your output assets. Required for all Nielsen non-linear watermarking. This ID should be unique to your Nielsen account but common to all of your output assets. Required for all Nielsen non-linear watermarking.

$sel:cbetSourceId:NielsenNonLinearWatermarkSettings', nielsenNonLinearWatermarkSettings_cbetSourceId - Use the CSID that Nielsen provides to you. This CBET source ID should be unique to your Nielsen account but common to all of your output assets that have CBET watermarking. Required when you choose a value for the setting Watermark types (ActiveWatermarkProcess) that includes CBET.

$sel:ticServerUrl:NielsenNonLinearWatermarkSettings', nielsenNonLinearWatermarkSettings_ticServerUrl - Specify the endpoint for the TIC server that you have deployed and configured in the AWS Cloud. Required for all Nielsen non-linear watermarking. MediaConvert can't connect directly to a TIC server. Instead, you must use API Gateway to provide a RESTful interface between MediaConvert and a TIC server that you deploy in your AWS account. For more information on deploying a TIC server in your AWS account and the required API Gateway, contact Nielsen support.

$sel:metadataDestination:NielsenNonLinearWatermarkSettings', nielsenNonLinearWatermarkSettings_metadataDestination - Specify the Amazon S3 location where you want MediaConvert to save your Nielsen non-linear metadata .zip file. This Amazon S3 bucket must be in the same Region as the one where you do your MediaConvert transcoding. If you want to include an ADI file in this .zip file, use the setting ADI file (adiFilename) to specify it. MediaConvert delivers the Nielsen metadata .zip files only to your metadata destination Amazon S3 bucket. It doesn't deliver the .zip files to Nielsen. You are responsible for delivering the metadata .zip files to Nielsen.

$sel:assetName:NielsenNonLinearWatermarkSettings', nielsenNonLinearWatermarkSettings_assetName - Use the asset name that you provide to Nielsen for this asset. Required for all Nielsen non-linear watermarking.

$sel:adiFilename:NielsenNonLinearWatermarkSettings', nielsenNonLinearWatermarkSettings_adiFilename - Optional. Use this setting when you want the service to include an ADI file in the Nielsen metadata .zip file. To provide an ADI file, store it in Amazon S3 and provide a URL to it here. The URL should be in the following format: S3://bucket/path/ADI-file. For more information about the metadata .zip file, see the setting Metadata destination (metadataDestination).

$sel:assetId:NielsenNonLinearWatermarkSettings', nielsenNonLinearWatermarkSettings_assetId - Use the asset ID that you provide to Nielsen to uniquely identify this asset. Required for all Nielsen non-linear watermarking.

$sel:uniqueTicPerAudioTrack:NielsenNonLinearWatermarkSettings', nielsenNonLinearWatermarkSettings_uniqueTicPerAudioTrack - To create assets that have the same TIC values in each audio track, keep the default value Share TICs (SAME_TICS_PER_TRACK). To create assets that have unique TIC values for each audio track, choose Use unique TICs (RESERVE_UNIQUE_TICS_PER_TRACK).

$sel:sourceWatermarkStatus:NielsenNonLinearWatermarkSettings', nielsenNonLinearWatermarkSettings_sourceWatermarkStatus - Required. Specify whether your source content already contains Nielsen non-linear watermarks. When you set this value to Watermarked (WATERMARKED), the service fails the job. Nielsen requires that you add non-linear watermarking to only clean content that doesn't already have non-linear Nielsen watermarks.

nielsenNonLinearWatermarkSettings_episodeId :: Lens' NielsenNonLinearWatermarkSettings (Maybe Text)

Optional. If this asset uses an episode ID with Nielsen, provide it here.

nielsenNonLinearWatermarkSettings_activeWatermarkProcess :: Lens' NielsenNonLinearWatermarkSettings (Maybe NielsenActiveWatermarkProcessType)

Choose the type of Nielsen watermarks that you want in your outputs. When you choose NAES 2 and NW (NAES2_AND_NW), you must provide a value for the setting SID (sourceId). When you choose CBET (CBET), you must provide a value for the setting CSID (cbetSourceId). When you choose NAES 2, NW, and CBET (NAES2_AND_NW_AND_CBET), you must provide values for both of these settings.

nielsenNonLinearWatermarkSettings_sourceId :: Lens' NielsenNonLinearWatermarkSettings (Maybe Natural)

Use the SID that Nielsen provides to you. This source ID should be unique to your Nielsen account but common to all of your output assets. Required for all Nielsen non-linear watermarking. This ID should be unique to your Nielsen account but common to all of your output assets. Required for all Nielsen non-linear watermarking.

nielsenNonLinearWatermarkSettings_cbetSourceId :: Lens' NielsenNonLinearWatermarkSettings (Maybe Text)

Use the CSID that Nielsen provides to you. This CBET source ID should be unique to your Nielsen account but common to all of your output assets that have CBET watermarking. Required when you choose a value for the setting Watermark types (ActiveWatermarkProcess) that includes CBET.

nielsenNonLinearWatermarkSettings_ticServerUrl :: Lens' NielsenNonLinearWatermarkSettings (Maybe Text)

Specify the endpoint for the TIC server that you have deployed and configured in the AWS Cloud. Required for all Nielsen non-linear watermarking. MediaConvert can't connect directly to a TIC server. Instead, you must use API Gateway to provide a RESTful interface between MediaConvert and a TIC server that you deploy in your AWS account. For more information on deploying a TIC server in your AWS account and the required API Gateway, contact Nielsen support.

nielsenNonLinearWatermarkSettings_metadataDestination :: Lens' NielsenNonLinearWatermarkSettings (Maybe Text)

Specify the Amazon S3 location where you want MediaConvert to save your Nielsen non-linear metadata .zip file. This Amazon S3 bucket must be in the same Region as the one where you do your MediaConvert transcoding. If you want to include an ADI file in this .zip file, use the setting ADI file (adiFilename) to specify it. MediaConvert delivers the Nielsen metadata .zip files only to your metadata destination Amazon S3 bucket. It doesn't deliver the .zip files to Nielsen. You are responsible for delivering the metadata .zip files to Nielsen.

nielsenNonLinearWatermarkSettings_assetName :: Lens' NielsenNonLinearWatermarkSettings (Maybe Text)

Use the asset name that you provide to Nielsen for this asset. Required for all Nielsen non-linear watermarking.

nielsenNonLinearWatermarkSettings_adiFilename :: Lens' NielsenNonLinearWatermarkSettings (Maybe Text)

Optional. Use this setting when you want the service to include an ADI file in the Nielsen metadata .zip file. To provide an ADI file, store it in Amazon S3 and provide a URL to it here. The URL should be in the following format: S3://bucket/path/ADI-file. For more information about the metadata .zip file, see the setting Metadata destination (metadataDestination).

nielsenNonLinearWatermarkSettings_assetId :: Lens' NielsenNonLinearWatermarkSettings (Maybe Text)

Use the asset ID that you provide to Nielsen to uniquely identify this asset. Required for all Nielsen non-linear watermarking.

nielsenNonLinearWatermarkSettings_uniqueTicPerAudioTrack :: Lens' NielsenNonLinearWatermarkSettings (Maybe NielsenUniqueTicPerAudioTrackType)

To create assets that have the same TIC values in each audio track, keep the default value Share TICs (SAME_TICS_PER_TRACK). To create assets that have unique TIC values for each audio track, choose Use unique TICs (RESERVE_UNIQUE_TICS_PER_TRACK).

nielsenNonLinearWatermarkSettings_sourceWatermarkStatus :: Lens' NielsenNonLinearWatermarkSettings (Maybe NielsenSourceWatermarkStatusType)

Required. Specify whether your source content already contains Nielsen non-linear watermarks. When you set this value to Watermarked (WATERMARKED), the service fails the job. Nielsen requires that you add non-linear watermarking to only clean content that doesn't already have non-linear Nielsen watermarks.