Network.AWS.MediaConvert.Types.NoiseReducerSpatialFilterSettings

Description

Documentation

data NoiseReducerSpatialFilterSettings

Noise reducer filter settings for spatial filter.

See: newNoiseReducerSpatialFilterSettings smart constructor.

Constructors

NoiseReducerSpatialFilterSettings'

Fields

strength :: Maybe Natural

Relative strength of noise reducing filter. Higher values produce stronger filtering.

postFilterSharpenStrength :: Maybe Natural

Specify strength of post noise reduction sharpening filter, with 0 disabling the filter and 3 enabling it at maximum strength.

speed :: Maybe Int

The speed of the filter, from -2 (lower speed) to 3 (higher speed), with 0 being the nominal value.

newNoiseReducerSpatialFilterSettings :: NoiseReducerSpatialFilterSettings

Create a value of NoiseReducerSpatialFilterSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:strength:NoiseReducerSpatialFilterSettings', noiseReducerSpatialFilterSettings_strength - Relative strength of noise reducing filter. Higher values produce stronger filtering.

$sel:postFilterSharpenStrength:NoiseReducerSpatialFilterSettings', noiseReducerSpatialFilterSettings_postFilterSharpenStrength - Specify strength of post noise reduction sharpening filter, with 0 disabling the filter and 3 enabling it at maximum strength.

$sel:speed:NoiseReducerSpatialFilterSettings', noiseReducerSpatialFilterSettings_speed - The speed of the filter, from -2 (lower speed) to 3 (higher speed), with 0 being the nominal value.

noiseReducerSpatialFilterSettings_strength :: Lens' NoiseReducerSpatialFilterSettings (Maybe Natural)

Relative strength of noise reducing filter. Higher values produce stronger filtering.

noiseReducerSpatialFilterSettings_postFilterSharpenStrength :: Lens' NoiseReducerSpatialFilterSettings (Maybe Natural)

Specify strength of post noise reduction sharpening filter, with 0 disabling the filter and 3 enabling it at maximum strength.

noiseReducerSpatialFilterSettings_speed :: Lens' NoiseReducerSpatialFilterSettings (Maybe Int)

The speed of the filter, from -2 (lower speed) to 3 (higher speed), with 0 being the nominal value.