Network.AWS.MediaConvert.Types.NoiseReducerTemporalFilterSettings

Description

Documentation

data NoiseReducerTemporalFilterSettings

Noise reducer filter settings for temporal filter.

See: newNoiseReducerTemporalFilterSettings smart constructor.

Constructors

NoiseReducerTemporalFilterSettings'

Fields

speed :: Maybe Int

The speed of the filter (higher number is faster). Low setting reduces bit rate at the cost of transcode time, high setting improves transcode time at the cost of bit rate.

aggressiveMode :: Maybe Natural

Use Aggressive mode for content that has complex motion. Higher values produce stronger temporal filtering. This filters highly complex scenes more aggressively and creates better VQ for low bitrate outputs.

postTemporalSharpening :: Maybe NoiseFilterPostTemporalSharpening

Optional. When you set Noise reducer (noiseReducer) to Temporal (TEMPORAL), you can use this setting to apply sharpening. The default behavior, Auto (AUTO), allows the transcoder to determine whether to apply filtering, depending on input type and quality. When you set Noise reducer to Temporal, your output bandwidth is reduced. When Post temporal sharpening is also enabled, that bandwidth reduction is smaller.

strength :: Maybe Natural

Specify the strength of the noise reducing filter on this output. Higher values produce stronger filtering. We recommend the following value ranges, depending on the result that you want: * 0-2 for complexity reduction with minimal sharpness loss * 2-8 for complexity reduction with image preservation * 8-16 for a high level of complexity reduction

newNoiseReducerTemporalFilterSettings :: NoiseReducerTemporalFilterSettings

Create a value of NoiseReducerTemporalFilterSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:speed:NoiseReducerTemporalFilterSettings', noiseReducerTemporalFilterSettings_speed - The speed of the filter (higher number is faster). Low setting reduces bit rate at the cost of transcode time, high setting improves transcode time at the cost of bit rate.

$sel:aggressiveMode:NoiseReducerTemporalFilterSettings', noiseReducerTemporalFilterSettings_aggressiveMode - Use Aggressive mode for content that has complex motion. Higher values produce stronger temporal filtering. This filters highly complex scenes more aggressively and creates better VQ for low bitrate outputs.

$sel:postTemporalSharpening:NoiseReducerTemporalFilterSettings', noiseReducerTemporalFilterSettings_postTemporalSharpening - Optional. When you set Noise reducer (noiseReducer) to Temporal (TEMPORAL), you can use this setting to apply sharpening. The default behavior, Auto (AUTO), allows the transcoder to determine whether to apply filtering, depending on input type and quality. When you set Noise reducer to Temporal, your output bandwidth is reduced. When Post temporal sharpening is also enabled, that bandwidth reduction is smaller.

$sel:strength:NoiseReducerTemporalFilterSettings', noiseReducerTemporalFilterSettings_strength - Specify the strength of the noise reducing filter on this output. Higher values produce stronger filtering. We recommend the following value ranges, depending on the result that you want: * 0-2 for complexity reduction with minimal sharpness loss * 2-8 for complexity reduction with image preservation * 8-16 for a high level of complexity reduction

noiseReducerTemporalFilterSettings_speed :: Lens' NoiseReducerTemporalFilterSettings (Maybe Int)

The speed of the filter (higher number is faster). Low setting reduces bit rate at the cost of transcode time, high setting improves transcode time at the cost of bit rate.

noiseReducerTemporalFilterSettings_aggressiveMode :: Lens' NoiseReducerTemporalFilterSettings (Maybe Natural)

Use Aggressive mode for content that has complex motion. Higher values produce stronger temporal filtering. This filters highly complex scenes more aggressively and creates better VQ for low bitrate outputs.

noiseReducerTemporalFilterSettings_postTemporalSharpening :: Lens' NoiseReducerTemporalFilterSettings (Maybe NoiseFilterPostTemporalSharpening)

Optional. When you set Noise reducer (noiseReducer) to Temporal (TEMPORAL), you can use this setting to apply sharpening. The default behavior, Auto (AUTO), allows the transcoder to determine whether to apply filtering, depending on input type and quality. When you set Noise reducer to Temporal, your output bandwidth is reduced. When Post temporal sharpening is also enabled, that bandwidth reduction is smaller.

noiseReducerTemporalFilterSettings_strength :: Lens' NoiseReducerTemporalFilterSettings (Maybe Natural)

Specify the strength of the noise reducing filter on this output. Higher values produce stronger filtering. We recommend the following value ranges, depending on the result that you want: * 0-2 for complexity reduction with minimal sharpness loss * 2-8 for complexity reduction with image preservation * 8-16 for a high level of complexity reduction