Network.AWS.MediaConvert.Types.OutputChannelMapping

Description

Documentation

data OutputChannelMapping

OutputChannel mapping settings.

See: newOutputChannelMapping smart constructor.

Constructors

OutputChannelMapping'

Fields

inputChannels :: Maybe [Int]

Use this setting to specify your remix values when they are integers, such as -10, 0, or 4.

inputChannelsFineTune :: Maybe [Double]

Use this setting to specify your remix values when they have a decimal component, such as -10.312, 0.08, or 4.9. MediaConvert rounds your remixing values to the nearest thousandth.

newOutputChannelMapping :: OutputChannelMapping

Create a value of OutputChannelMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputChannels:OutputChannelMapping', outputChannelMapping_inputChannels - Use this setting to specify your remix values when they are integers, such as -10, 0, or 4.

$sel:inputChannelsFineTune:OutputChannelMapping', outputChannelMapping_inputChannelsFineTune - Use this setting to specify your remix values when they have a decimal component, such as -10.312, 0.08, or 4.9. MediaConvert rounds your remixing values to the nearest thousandth.

outputChannelMapping_inputChannels :: Lens' OutputChannelMapping (Maybe [Int])

Use this setting to specify your remix values when they are integers, such as -10, 0, or 4.

outputChannelMapping_inputChannelsFineTune :: Lens' OutputChannelMapping (Maybe [Double])

Use this setting to specify your remix values when they have a decimal component, such as -10.312, 0.08, or 4.9. MediaConvert rounds your remixing values to the nearest thousandth.