Network.AWS.MediaConvert.Types.OutputGroupSettings

Description

Documentation

data OutputGroupSettings

Output Group settings, including type

See: newOutputGroupSettings smart constructor.

Constructors

OutputGroupSettings'

Fields

fileGroupSettings :: Maybe FileGroupSettings

Settings related to your File output group. MediaConvert uses this group of settings to generate a single standalone file, rather than a streaming package. When you work directly in your JSON job specification, include this object and any required children when you set Type, under OutputGroupSettings, to FILE_GROUP_SETTINGS.

cmafGroupSettings :: Maybe CmafGroupSettings

Settings related to your CMAF output package. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/outputs-file-ABR.html. When you work directly in your JSON job specification, include this object and any required children when you set Type, under OutputGroupSettings, to CMAF_GROUP_SETTINGS.

msSmoothGroupSettings :: Maybe MsSmoothGroupSettings

Settings related to your Microsoft Smooth Streaming output package. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/outputs-file-ABR.html. When you work directly in your JSON job specification, include this object and any required children when you set Type, under OutputGroupSettings, to MS_SMOOTH_GROUP_SETTINGS.

hlsGroupSettings :: Maybe HlsGroupSettings

Settings related to your HLS output package. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/outputs-file-ABR.html. When you work directly in your JSON job specification, include this object and any required children when you set Type, under OutputGroupSettings, to HLS_GROUP_SETTINGS.

type' :: Maybe OutputGroupType

Type of output group (File group, Apple HLS, DASH ISO, Microsoft Smooth Streaming, CMAF)

dashIsoGroupSettings :: Maybe DashIsoGroupSettings

Settings related to your DASH output package. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/outputs-file-ABR.html. When you work directly in your JSON job specification, include this object and any required children when you set Type, under OutputGroupSettings, to DASH_ISO_GROUP_SETTINGS.

newOutputGroupSettings :: OutputGroupSettings

Create a value of OutputGroupSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:fileGroupSettings:OutputGroupSettings', outputGroupSettings_fileGroupSettings - Settings related to your File output group. MediaConvert uses this group of settings to generate a single standalone file, rather than a streaming package. When you work directly in your JSON job specification, include this object and any required children when you set Type, under OutputGroupSettings, to FILE_GROUP_SETTINGS.

$sel:cmafGroupSettings:OutputGroupSettings', outputGroupSettings_cmafGroupSettings - Settings related to your CMAF output package. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/outputs-file-ABR.html. When you work directly in your JSON job specification, include this object and any required children when you set Type, under OutputGroupSettings, to CMAF_GROUP_SETTINGS.

$sel:msSmoothGroupSettings:OutputGroupSettings', outputGroupSettings_msSmoothGroupSettings - Settings related to your Microsoft Smooth Streaming output package. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/outputs-file-ABR.html. When you work directly in your JSON job specification, include this object and any required children when you set Type, under OutputGroupSettings, to MS_SMOOTH_GROUP_SETTINGS.

$sel:hlsGroupSettings:OutputGroupSettings', outputGroupSettings_hlsGroupSettings - Settings related to your HLS output package. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/outputs-file-ABR.html. When you work directly in your JSON job specification, include this object and any required children when you set Type, under OutputGroupSettings, to HLS_GROUP_SETTINGS.

$sel:type':OutputGroupSettings', outputGroupSettings_type - Type of output group (File group, Apple HLS, DASH ISO, Microsoft Smooth Streaming, CMAF)

$sel:dashIsoGroupSettings:OutputGroupSettings', outputGroupSettings_dashIsoGroupSettings - Settings related to your DASH output package. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/outputs-file-ABR.html. When you work directly in your JSON job specification, include this object and any required children when you set Type, under OutputGroupSettings, to DASH_ISO_GROUP_SETTINGS.

outputGroupSettings_fileGroupSettings :: Lens' OutputGroupSettings (Maybe FileGroupSettings)

Settings related to your File output group. MediaConvert uses this group of settings to generate a single standalone file, rather than a streaming package. When you work directly in your JSON job specification, include this object and any required children when you set Type, under OutputGroupSettings, to FILE_GROUP_SETTINGS.

outputGroupSettings_cmafGroupSettings :: Lens' OutputGroupSettings (Maybe CmafGroupSettings)

Settings related to your CMAF output package. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/outputs-file-ABR.html. When you work directly in your JSON job specification, include this object and any required children when you set Type, under OutputGroupSettings, to CMAF_GROUP_SETTINGS.

outputGroupSettings_msSmoothGroupSettings :: Lens' OutputGroupSettings (Maybe MsSmoothGroupSettings)

Settings related to your Microsoft Smooth Streaming output package. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/outputs-file-ABR.html. When you work directly in your JSON job specification, include this object and any required children when you set Type, under OutputGroupSettings, to MS_SMOOTH_GROUP_SETTINGS.

outputGroupSettings_hlsGroupSettings :: Lens' OutputGroupSettings (Maybe HlsGroupSettings)

Settings related to your HLS output package. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/outputs-file-ABR.html. When you work directly in your JSON job specification, include this object and any required children when you set Type, under OutputGroupSettings, to HLS_GROUP_SETTINGS.

outputGroupSettings_type :: Lens' OutputGroupSettings (Maybe OutputGroupType)

Type of output group (File group, Apple HLS, DASH ISO, Microsoft Smooth Streaming, CMAF)

outputGroupSettings_dashIsoGroupSettings :: Lens' OutputGroupSettings (Maybe DashIsoGroupSettings)

Settings related to your DASH output package. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/outputs-file-ABR.html. When you work directly in your JSON job specification, include this object and any required children when you set Type, under OutputGroupSettings, to DASH_ISO_GROUP_SETTINGS.