Network.AWS.MediaConvert.Types.PartnerWatermarking

Description

Documentation

data PartnerWatermarking

If you work with a third party video watermarking partner, use the group of settings that correspond with your watermarking partner to include watermarks in your output.

See: newPartnerWatermarking smart constructor.

Constructors

PartnerWatermarking'

Fields

nexguardFileMarkerSettings :: Maybe NexGuardFileMarkerSettings

For forensic video watermarking, MediaConvert supports Nagra NexGuard File Marker watermarking. MediaConvert supports both PreRelease Content (NGPR/G2) and OTT Streaming workflows.

newPartnerWatermarking :: PartnerWatermarking

Create a value of PartnerWatermarking with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nexguardFileMarkerSettings:PartnerWatermarking', partnerWatermarking_nexguardFileMarkerSettings - For forensic video watermarking, MediaConvert supports Nagra NexGuard File Marker watermarking. MediaConvert supports both PreRelease Content (NGPR/G2) and OTT Streaming workflows.

partnerWatermarking_nexguardFileMarkerSettings :: Lens' PartnerWatermarking (Maybe NexGuardFileMarkerSettings)

For forensic video watermarking, MediaConvert supports Nagra NexGuard File Marker watermarking. MediaConvert supports both PreRelease Content (NGPR/G2) and OTT Streaming workflows.