Network.AWS.MediaConvert.Types.PresetSettings

Description

Documentation

data PresetSettings

Settings for preset

See: newPresetSettings smart constructor.

Constructors

PresetSettings'

Fields

captionDescriptions :: Maybe [CaptionDescriptionPreset]

This object holds groups of settings related to captions for one output. For each output that has captions, include one instance of CaptionDescriptions.

videoDescription :: Maybe VideoDescription

VideoDescription contains a group of video encoding settings. The specific video settings depend on the video codec that you choose for the property codec. Include one instance of VideoDescription per output.

containerSettings :: Maybe ContainerSettings

Container specific settings.

audioDescriptions :: Maybe [AudioDescription]

(AudioDescriptions) contains groups of audio encoding settings organized by audio codec. Include one instance of (AudioDescriptions) per output. (AudioDescriptions) can contain multiple groups of encoding settings.

newPresetSettings :: PresetSettings

Create a value of PresetSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:captionDescriptions:PresetSettings', presetSettings_captionDescriptions - This object holds groups of settings related to captions for one output. For each output that has captions, include one instance of CaptionDescriptions.

$sel:videoDescription:PresetSettings', presetSettings_videoDescription - VideoDescription contains a group of video encoding settings. The specific video settings depend on the video codec that you choose for the property codec. Include one instance of VideoDescription per output.

$sel:containerSettings:PresetSettings', presetSettings_containerSettings - Container specific settings.

$sel:audioDescriptions:PresetSettings', presetSettings_audioDescriptions - (AudioDescriptions) contains groups of audio encoding settings organized by audio codec. Include one instance of (AudioDescriptions) per output. (AudioDescriptions) can contain multiple groups of encoding settings.

presetSettings_captionDescriptions :: Lens' PresetSettings (Maybe [CaptionDescriptionPreset])

This object holds groups of settings related to captions for one output. For each output that has captions, include one instance of CaptionDescriptions.

presetSettings_videoDescription :: Lens' PresetSettings (Maybe VideoDescription)

VideoDescription contains a group of video encoding settings. The specific video settings depend on the video codec that you choose for the property codec. Include one instance of VideoDescription per output.

presetSettings_containerSettings :: Lens' PresetSettings (Maybe ContainerSettings)

Container specific settings.

presetSettings_audioDescriptions :: Lens' PresetSettings (Maybe [AudioDescription])

(AudioDescriptions) contains groups of audio encoding settings organized by audio codec. Include one instance of (AudioDescriptions) per output. (AudioDescriptions) can contain multiple groups of encoding settings.