Network.AWS.MediaConvert.Types.Queue

Description

Documentation

data Queue

You can use queues to manage the resources that are available to your AWS account for running multiple transcoding jobs at the same time. If you don't specify a queue, the service sends all jobs through the default queue. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/working-with-queues.html.

See: newQueue smart constructor.

Constructors

Queue'

Fields

status :: Maybe QueueStatus

Queues can be ACTIVE or PAUSED. If you pause a queue, the service won't begin processing jobs in that queue. Jobs that are running when you pause the queue continue to run until they finish or result in an error.

createdAt :: Maybe POSIX

The timestamp in epoch seconds for when you created the queue.

arn :: Maybe Text

An identifier for this resource that is unique within all of AWS.

lastUpdated :: Maybe POSIX

The timestamp in epoch seconds for when you most recently updated the queue.

description :: Maybe Text

An optional description that you create for each queue.

submittedJobsCount :: Maybe Int

The estimated number of jobs with a SUBMITTED status.

reservationPlan :: Maybe ReservationPlan

Details about the pricing plan for your reserved queue. Required for reserved queues and not applicable to on-demand queues.

pricingPlan :: Maybe PricingPlan

Specifies whether the pricing plan for the queue is on-demand or reserved. For on-demand, you pay per minute, billed in increments of .01 minute. For reserved, you pay for the transcoding capacity of the entire queue, regardless of how much or how little you use it. Reserved pricing requires a 12-month commitment.

type' :: Maybe Type

Specifies whether this on-demand queue is system or custom. System queues are built in. You can't modify or delete system queues. You can create and modify custom queues.

progressingJobsCount :: Maybe Int

The estimated number of jobs with a PROGRESSING status.

name :: Text

A name that you create for each queue. Each name must be unique within your account.

newQueue

Arguments

:: Text
$sel:name:Queue'

-> Queue

Create a value of Queue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Queue', queue_status - Queues can be ACTIVE or PAUSED. If you pause a queue, the service won't begin processing jobs in that queue. Jobs that are running when you pause the queue continue to run until they finish or result in an error.

$sel:createdAt:Queue', queue_createdAt - The timestamp in epoch seconds for when you created the queue.

$sel:arn:Queue', queue_arn - An identifier for this resource that is unique within all of AWS.

$sel:lastUpdated:Queue', queue_lastUpdated - The timestamp in epoch seconds for when you most recently updated the queue.

$sel:description:Queue', queue_description - An optional description that you create for each queue.

$sel:submittedJobsCount:Queue', queue_submittedJobsCount - The estimated number of jobs with a SUBMITTED status.

$sel:reservationPlan:Queue', queue_reservationPlan - Details about the pricing plan for your reserved queue. Required for reserved queues and not applicable to on-demand queues.

$sel:pricingPlan:Queue', queue_pricingPlan - Specifies whether the pricing plan for the queue is on-demand or reserved. For on-demand, you pay per minute, billed in increments of .01 minute. For reserved, you pay for the transcoding capacity of the entire queue, regardless of how much or how little you use it. Reserved pricing requires a 12-month commitment.

$sel:type':Queue', queue_type - Specifies whether this on-demand queue is system or custom. System queues are built in. You can't modify or delete system queues. You can create and modify custom queues.

$sel:progressingJobsCount:Queue', queue_progressingJobsCount - The estimated number of jobs with a PROGRESSING status.

$sel:name:Queue', queue_name - A name that you create for each queue. Each name must be unique within your account.

queue_status :: Lens' Queue (Maybe QueueStatus)

Queues can be ACTIVE or PAUSED. If you pause a queue, the service won't begin processing jobs in that queue. Jobs that are running when you pause the queue continue to run until they finish or result in an error.

queue_createdAt :: Lens' Queue (Maybe UTCTime)

The timestamp in epoch seconds for when you created the queue.

queue_arn :: Lens' Queue (Maybe Text)

An identifier for this resource that is unique within all of AWS.

queue_lastUpdated :: Lens' Queue (Maybe UTCTime)

The timestamp in epoch seconds for when you most recently updated the queue.

queue_description :: Lens' Queue (Maybe Text)

An optional description that you create for each queue.

queue_submittedJobsCount :: Lens' Queue (Maybe Int)

The estimated number of jobs with a SUBMITTED status.

queue_reservationPlan :: Lens' Queue (Maybe ReservationPlan)

Details about the pricing plan for your reserved queue. Required for reserved queues and not applicable to on-demand queues.

queue_pricingPlan :: Lens' Queue (Maybe PricingPlan)

Specifies whether the pricing plan for the queue is on-demand or reserved. For on-demand, you pay per minute, billed in increments of .01 minute. For reserved, you pay for the transcoding capacity of the entire queue, regardless of how much or how little you use it. Reserved pricing requires a 12-month commitment.

queue_type :: Lens' Queue (Maybe Type)

Specifies whether this on-demand queue is system or custom. System queues are built in. You can't modify or delete system queues. You can create and modify custom queues.

queue_progressingJobsCount :: Lens' Queue (Maybe Int)

The estimated number of jobs with a PROGRESSING status.

queue_name :: Lens' Queue Text

A name that you create for each queue. Each name must be unique within your account.