Network.AWS.MediaConvert.Types.SpekeKeyProvider

Description

Documentation

data SpekeKeyProvider

If your output group type is HLS, DASH, or Microsoft Smooth, use these settings when doing DRM encryption with a SPEKE-compliant key provider. If your output group type is CMAF, use the SpekeKeyProviderCmaf settings instead.

See: newSpekeKeyProvider smart constructor.

Constructors

SpekeKeyProvider'

Fields

resourceId :: Maybe Text

Specify the resource ID that your SPEKE-compliant key provider uses to identify this content.

certificateArn :: Maybe Text

If you want your key provider to encrypt the content keys that it provides to MediaConvert, set up a certificate with a master key using AWS Certificate Manager. Specify the certificate's Amazon Resource Name (ARN) here.

url :: Maybe Text

Specify the URL to the key server that your SPEKE-compliant DRM key provider uses to provide keys for encrypting your content.

systemIds :: Maybe [Text]

Relates to SPEKE implementation. DRM system identifiers. DASH output groups support a max of two system ids. Other group types support one system id. See https://dashif.org/identifiers/content_protection/ for more details.

newSpekeKeyProvider :: SpekeKeyProvider

Create a value of SpekeKeyProvider with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:SpekeKeyProvider', spekeKeyProvider_resourceId - Specify the resource ID that your SPEKE-compliant key provider uses to identify this content.

$sel:certificateArn:SpekeKeyProvider', spekeKeyProvider_certificateArn - If you want your key provider to encrypt the content keys that it provides to MediaConvert, set up a certificate with a master key using AWS Certificate Manager. Specify the certificate's Amazon Resource Name (ARN) here.

$sel:url:SpekeKeyProvider', spekeKeyProvider_url - Specify the URL to the key server that your SPEKE-compliant DRM key provider uses to provide keys for encrypting your content.

$sel:systemIds:SpekeKeyProvider', spekeKeyProvider_systemIds - Relates to SPEKE implementation. DRM system identifiers. DASH output groups support a max of two system ids. Other group types support one system id. See https://dashif.org/identifiers/content_protection/ for more details.

spekeKeyProvider_resourceId :: Lens' SpekeKeyProvider (Maybe Text)

Specify the resource ID that your SPEKE-compliant key provider uses to identify this content.

spekeKeyProvider_certificateArn :: Lens' SpekeKeyProvider (Maybe Text)

If you want your key provider to encrypt the content keys that it provides to MediaConvert, set up a certificate with a master key using AWS Certificate Manager. Specify the certificate's Amazon Resource Name (ARN) here.

spekeKeyProvider_url :: Lens' SpekeKeyProvider (Maybe Text)

Specify the URL to the key server that your SPEKE-compliant DRM key provider uses to provide keys for encrypting your content.

spekeKeyProvider_systemIds :: Lens' SpekeKeyProvider (Maybe [Text])

Relates to SPEKE implementation. DRM system identifiers. DASH output groups support a max of two system ids. Other group types support one system id. See https://dashif.org/identifiers/content_protection/ for more details.