Network.AWS.MediaConvert.Types.TeletextDestinationSettings

Description

Documentation

data TeletextDestinationSettings

Settings related to teletext captions. Set up teletext captions in the same output as your video. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/teletext-output-captions.html. When you work directly in your JSON job specification, include this object and any required children when you set destinationType to TELETEXT.

See: newTeletextDestinationSettings smart constructor.

Constructors

TeletextDestinationSettings'

Fields

pageTypes :: Maybe [TeletextPageType]

Specify the page types for this Teletext page. If you don't specify a value here, the service sets the page type to the default value Subtitle (PAGE_TYPE_SUBTITLE). If you pass through the entire set of Teletext data, don't use this field. When you pass through a set of Teletext pages, your output has the same page types as your input.

pageNumber :: Maybe Text

Set pageNumber to the Teletext page number for the destination captions for this output. This value must be a three-digit hexadecimal string; strings ending in -FF are invalid. If you are passing through the entire set of Teletext data, do not use this field.

newTeletextDestinationSettings :: TeletextDestinationSettings

Create a value of TeletextDestinationSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pageTypes:TeletextDestinationSettings', teletextDestinationSettings_pageTypes - Specify the page types for this Teletext page. If you don't specify a value here, the service sets the page type to the default value Subtitle (PAGE_TYPE_SUBTITLE). If you pass through the entire set of Teletext data, don't use this field. When you pass through a set of Teletext pages, your output has the same page types as your input.

$sel:pageNumber:TeletextDestinationSettings', teletextDestinationSettings_pageNumber - Set pageNumber to the Teletext page number for the destination captions for this output. This value must be a three-digit hexadecimal string; strings ending in -FF are invalid. If you are passing through the entire set of Teletext data, do not use this field.

teletextDestinationSettings_pageTypes :: Lens' TeletextDestinationSettings (Maybe [TeletextPageType])

Specify the page types for this Teletext page. If you don't specify a value here, the service sets the page type to the default value Subtitle (PAGE_TYPE_SUBTITLE). If you pass through the entire set of Teletext data, don't use this field. When you pass through a set of Teletext pages, your output has the same page types as your input.

teletextDestinationSettings_pageNumber :: Lens' TeletextDestinationSettings (Maybe Text)

Set pageNumber to the Teletext page number for the destination captions for this output. This value must be a three-digit hexadecimal string; strings ending in -FF are invalid. If you are passing through the entire set of Teletext data, do not use this field.