Network.AWS.MediaConvert.Types.VideoCodecSettings

Description

Documentation

data VideoCodecSettings

Video codec settings, (CodecSettings) under (VideoDescription), contains the group of settings related to video encoding. The settings in this group vary depending on the value that you choose for Video codec (Codec). For each codec enum that you choose, define the corresponding settings object. The following lists the codec enum, settings object pairs. * AV1, Av1Settings * AVC_INTRA, AvcIntraSettings * FRAME_CAPTURE, FrameCaptureSettings * H_264, H264Settings * H_265, H265Settings * MPEG2, Mpeg2Settings * PRORES, ProresSettings * VC3, Vc3Settings * VP8, Vp8Settings * VP9, Vp9Settings * XAVC, XavcSettings

See: newVideoCodecSettings smart constructor.

Constructors

VideoCodecSettings'

Fields

frameCaptureSettings :: Maybe FrameCaptureSettings

Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value FRAME_CAPTURE.

xavcSettings :: Maybe XavcSettings

Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value XAVC.

codec :: Maybe VideoCodec

Specifies the video codec. This must be equal to one of the enum values defined by the object VideoCodec.

vc3Settings :: Maybe Vc3Settings

Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value VC3

vp8Settings :: Maybe Vp8Settings

Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value VP8.

mpeg2Settings :: Maybe Mpeg2Settings

Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value MPEG2.

h264Settings :: Maybe H264Settings

Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value H_264.

vp9Settings :: Maybe Vp9Settings

Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value VP9.

h265Settings :: Maybe H265Settings

Settings for H265 codec

proresSettings :: Maybe ProresSettings

Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value PRORES.

avcIntraSettings :: Maybe AvcIntraSettings

Required when you choose AVC-Intra for your output video codec. For more information about the AVC-Intra settings, see the relevant specification. For detailed information about SD and HD in AVC-Intra, see https://ieeexplore.ieee.org/document/7290936. For information about 4K/2K in AVC-Intra, see https://pro-av.panasonic.net/en/avc-ultra/AVC-ULTRAoverview.pdf.

av1Settings :: Maybe Av1Settings

Required when you set Codec, under VideoDescription>CodecSettings to the value AV1.

newVideoCodecSettings :: VideoCodecSettings

Create a value of VideoCodecSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:frameCaptureSettings:VideoCodecSettings', videoCodecSettings_frameCaptureSettings - Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value FRAME_CAPTURE.

$sel:xavcSettings:VideoCodecSettings', videoCodecSettings_xavcSettings - Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value XAVC.

$sel:codec:VideoCodecSettings', videoCodecSettings_codec - Specifies the video codec. This must be equal to one of the enum values defined by the object VideoCodec.

$sel:vc3Settings:VideoCodecSettings', videoCodecSettings_vc3Settings - Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value VC3

$sel:vp8Settings:VideoCodecSettings', videoCodecSettings_vp8Settings - Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value VP8.

$sel:mpeg2Settings:VideoCodecSettings', videoCodecSettings_mpeg2Settings - Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value MPEG2.

$sel:h264Settings:VideoCodecSettings', videoCodecSettings_h264Settings - Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value H_264.

$sel:vp9Settings:VideoCodecSettings', videoCodecSettings_vp9Settings - Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value VP9.

$sel:h265Settings:VideoCodecSettings', videoCodecSettings_h265Settings - Settings for H265 codec

$sel:proresSettings:VideoCodecSettings', videoCodecSettings_proresSettings - Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value PRORES.

$sel:avcIntraSettings:VideoCodecSettings', videoCodecSettings_avcIntraSettings - Required when you choose AVC-Intra for your output video codec. For more information about the AVC-Intra settings, see the relevant specification. For detailed information about SD and HD in AVC-Intra, see https://ieeexplore.ieee.org/document/7290936. For information about 4K/2K in AVC-Intra, see https://pro-av.panasonic.net/en/avc-ultra/AVC-ULTRAoverview.pdf.

$sel:av1Settings:VideoCodecSettings', videoCodecSettings_av1Settings - Required when you set Codec, under VideoDescription>CodecSettings to the value AV1.

videoCodecSettings_frameCaptureSettings :: Lens' VideoCodecSettings (Maybe FrameCaptureSettings)

Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value FRAME_CAPTURE.

videoCodecSettings_xavcSettings :: Lens' VideoCodecSettings (Maybe XavcSettings)

Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value XAVC.

videoCodecSettings_codec :: Lens' VideoCodecSettings (Maybe VideoCodec)

Specifies the video codec. This must be equal to one of the enum values defined by the object VideoCodec.

videoCodecSettings_vc3Settings :: Lens' VideoCodecSettings (Maybe Vc3Settings)

Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value VC3

videoCodecSettings_vp8Settings :: Lens' VideoCodecSettings (Maybe Vp8Settings)

Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value VP8.

videoCodecSettings_mpeg2Settings :: Lens' VideoCodecSettings (Maybe Mpeg2Settings)

Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value MPEG2.

videoCodecSettings_h264Settings :: Lens' VideoCodecSettings (Maybe H264Settings)

Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value H_264.

videoCodecSettings_vp9Settings :: Lens' VideoCodecSettings (Maybe Vp9Settings)

Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value VP9.

videoCodecSettings_h265Settings :: Lens' VideoCodecSettings (Maybe H265Settings)

Settings for H265 codec

videoCodecSettings_proresSettings :: Lens' VideoCodecSettings (Maybe ProresSettings)

Required when you set (Codec) under (VideoDescription)>(CodecSettings) to the value PRORES.

videoCodecSettings_avcIntraSettings :: Lens' VideoCodecSettings (Maybe AvcIntraSettings)

Required when you choose AVC-Intra for your output video codec. For more information about the AVC-Intra settings, see the relevant specification. For detailed information about SD and HD in AVC-Intra, see https://ieeexplore.ieee.org/document/7290936. For information about 4K/2K in AVC-Intra, see https://pro-av.panasonic.net/en/avc-ultra/AVC-ULTRAoverview.pdf.

videoCodecSettings_av1Settings :: Lens' VideoCodecSettings (Maybe Av1Settings)

Required when you set Codec, under VideoDescription>CodecSettings to the value AV1.