Network.AWS.MediaConvert.Types.VorbisSettings

Description

Documentation

data VorbisSettings

Required when you set Codec, under AudioDescriptions>CodecSettings, to the value Vorbis.

See: newVorbisSettings smart constructor.

Constructors

VorbisSettings'

Fields

channels :: Maybe Natural

Optional. Specify the number of channels in this output audio track. Choosing Mono on the console gives you 1 output channel; choosing Stereo gives you 2. In the API, valid values are 1 and 2. The default value is 2.

sampleRate :: Maybe Natural

Optional. Specify the audio sample rate in Hz. Valid values are 22050, 32000, 44100, and 48000. The default value is 48000.

vbrQuality :: Maybe Int

Optional. Specify the variable audio quality of this Vorbis output from -1 (lowest quality, ~45 kbit/s) to 10 (highest quality, ~500 kbit/s). The default value is 4 (~128 kbit/s). Values 5 and 6 are approximately 160 and 192 kbit/s, respectively.

newVorbisSettings :: VorbisSettings

Create a value of VorbisSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channels:VorbisSettings', vorbisSettings_channels - Optional. Specify the number of channels in this output audio track. Choosing Mono on the console gives you 1 output channel; choosing Stereo gives you 2. In the API, valid values are 1 and 2. The default value is 2.

$sel:sampleRate:VorbisSettings', vorbisSettings_sampleRate - Optional. Specify the audio sample rate in Hz. Valid values are 22050, 32000, 44100, and 48000. The default value is 48000.

$sel:vbrQuality:VorbisSettings', vorbisSettings_vbrQuality - Optional. Specify the variable audio quality of this Vorbis output from -1 (lowest quality, ~45 kbit/s) to 10 (highest quality, ~500 kbit/s). The default value is 4 (~128 kbit/s). Values 5 and 6 are approximately 160 and 192 kbit/s, respectively.

vorbisSettings_channels :: Lens' VorbisSettings (Maybe Natural)

Optional. Specify the number of channels in this output audio track. Choosing Mono on the console gives you 1 output channel; choosing Stereo gives you 2. In the API, valid values are 1 and 2. The default value is 2.

vorbisSettings_sampleRate :: Lens' VorbisSettings (Maybe Natural)

Optional. Specify the audio sample rate in Hz. Valid values are 22050, 32000, 44100, and 48000. The default value is 48000.

vorbisSettings_vbrQuality :: Lens' VorbisSettings (Maybe Int)

Optional. Specify the variable audio quality of this Vorbis output from -1 (lowest quality, ~45 kbit/s) to 10 (highest quality, ~500 kbit/s). The default value is 4 (~128 kbit/s). Values 5 and 6 are approximately 160 and 192 kbit/s, respectively.