Network.AWS.MediaConvert.Types.WavSettings

Description

Documentation

data WavSettings

Required when you set (Codec) under (AudioDescriptions)>(CodecSettings) to the value WAV.

See: newWavSettings smart constructor.

Constructors

WavSettings'

Fields

format :: Maybe WavFormat

The service defaults to using RIFF for WAV outputs. If your output audio is likely to exceed 4 GB in file size, or if you otherwise need the extended support of the RF64 format, set your output WAV file format to RF64.

channels :: Maybe Natural

Specify the number of channels in this output audio track. Valid values are 1 and even numbers up to 64. For example, 1, 2, 4, 6, and so on, up to 64.

bitDepth :: Maybe Natural

Specify Bit depth (BitDepth), in bits per sample, to choose the encoding quality for this audio track.

sampleRate :: Maybe Natural

Sample rate in Hz.

newWavSettings :: WavSettings

Create a value of WavSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:format:WavSettings', wavSettings_format - The service defaults to using RIFF for WAV outputs. If your output audio is likely to exceed 4 GB in file size, or if you otherwise need the extended support of the RF64 format, set your output WAV file format to RF64.

$sel:channels:WavSettings', wavSettings_channels - Specify the number of channels in this output audio track. Valid values are 1 and even numbers up to 64. For example, 1, 2, 4, 6, and so on, up to 64.

$sel:bitDepth:WavSettings', wavSettings_bitDepth - Specify Bit depth (BitDepth), in bits per sample, to choose the encoding quality for this audio track.

$sel:sampleRate:WavSettings', wavSettings_sampleRate - Sample rate in Hz.

wavSettings_format :: Lens' WavSettings (Maybe WavFormat)

The service defaults to using RIFF for WAV outputs. If your output audio is likely to exceed 4 GB in file size, or if you otherwise need the extended support of the RF64 format, set your output WAV file format to RF64.

wavSettings_channels :: Lens' WavSettings (Maybe Natural)

Specify the number of channels in this output audio track. Valid values are 1 and even numbers up to 64. For example, 1, 2, 4, 6, and so on, up to 64.

wavSettings_bitDepth :: Lens' WavSettings (Maybe Natural)

Specify Bit depth (BitDepth), in bits per sample, to choose the encoding quality for this audio track.

wavSettings_sampleRate :: Lens' WavSettings (Maybe Natural)

Sample rate in Hz.