Network.AWS.MediaConvert.UpdateJobTemplate

Description

Modify one of your existing job templates.

Creating a Request

data UpdateJobTemplate

See: newUpdateJobTemplate smart constructor.

Constructors

UpdateJobTemplate'

Fields

accelerationSettings :: Maybe AccelerationSettings

Accelerated transcoding can significantly speed up jobs with long, visually complex content. Outputs that use this feature incur pro-tier pricing. For information about feature limitations, see the AWS Elemental MediaConvert User Guide.

priority :: Maybe Int

Specify the relative priority for this job. In any given queue, the service begins processing the job with the highest value first. When more than one job has the same priority, the service begins processing the job that you submitted first. If you don't specify a priority, the service uses the default value 0.

statusUpdateInterval :: Maybe StatusUpdateInterval

Specify how often MediaConvert sends STATUS_UPDATE events to Amazon CloudWatch Events. Set the interval, in seconds, between status updates. MediaConvert sends an update at this interval from the time the service begins processing your job to the time it completes the transcode or encounters an error.

settings :: Maybe JobTemplateSettings

JobTemplateSettings contains all the transcode settings saved in the template that will be applied to jobs created from it.

category :: Maybe Text

The new category for the job template, if you are changing it.

hopDestinations :: Maybe [HopDestination]

Optional list of hop destinations.

queue :: Maybe Text

The new queue for the job template, if you are changing it.

description :: Maybe Text

The new description for the job template, if you are changing it.

name :: Text

The name of the job template you are modifying

newUpdateJobTemplate

Arguments

:: Text
$sel:name:UpdateJobTemplate'

-> UpdateJobTemplate

Create a value of UpdateJobTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accelerationSettings:UpdateJobTemplate', updateJobTemplate_accelerationSettings - Accelerated transcoding can significantly speed up jobs with long, visually complex content. Outputs that use this feature incur pro-tier pricing. For information about feature limitations, see the AWS Elemental MediaConvert User Guide.

$sel:priority:UpdateJobTemplate', updateJobTemplate_priority - Specify the relative priority for this job. In any given queue, the service begins processing the job with the highest value first. When more than one job has the same priority, the service begins processing the job that you submitted first. If you don't specify a priority, the service uses the default value 0.

$sel:statusUpdateInterval:UpdateJobTemplate', updateJobTemplate_statusUpdateInterval - Specify how often MediaConvert sends STATUS_UPDATE events to Amazon CloudWatch Events. Set the interval, in seconds, between status updates. MediaConvert sends an update at this interval from the time the service begins processing your job to the time it completes the transcode or encounters an error.

$sel:settings:UpdateJobTemplate', updateJobTemplate_settings - JobTemplateSettings contains all the transcode settings saved in the template that will be applied to jobs created from it.

$sel:category:UpdateJobTemplate', updateJobTemplate_category - The new category for the job template, if you are changing it.

$sel:hopDestinations:UpdateJobTemplate', updateJobTemplate_hopDestinations - Optional list of hop destinations.

$sel:queue:UpdateJobTemplate', updateJobTemplate_queue - The new queue for the job template, if you are changing it.

$sel:description:UpdateJobTemplate', updateJobTemplate_description - The new description for the job template, if you are changing it.

$sel:name:UpdateJobTemplate', updateJobTemplate_name - The name of the job template you are modifying

Request Lenses

updateJobTemplate_accelerationSettings :: Lens' UpdateJobTemplate (Maybe AccelerationSettings)

Accelerated transcoding can significantly speed up jobs with long, visually complex content. Outputs that use this feature incur pro-tier pricing. For information about feature limitations, see the AWS Elemental MediaConvert User Guide.

updateJobTemplate_priority :: Lens' UpdateJobTemplate (Maybe Int)

Specify the relative priority for this job. In any given queue, the service begins processing the job with the highest value first. When more than one job has the same priority, the service begins processing the job that you submitted first. If you don't specify a priority, the service uses the default value 0.

updateJobTemplate_statusUpdateInterval :: Lens' UpdateJobTemplate (Maybe StatusUpdateInterval)

Specify how often MediaConvert sends STATUS_UPDATE events to Amazon CloudWatch Events. Set the interval, in seconds, between status updates. MediaConvert sends an update at this interval from the time the service begins processing your job to the time it completes the transcode or encounters an error.

updateJobTemplate_settings :: Lens' UpdateJobTemplate (Maybe JobTemplateSettings)

JobTemplateSettings contains all the transcode settings saved in the template that will be applied to jobs created from it.

updateJobTemplate_category :: Lens' UpdateJobTemplate (Maybe Text)

The new category for the job template, if you are changing it.

updateJobTemplate_hopDestinations :: Lens' UpdateJobTemplate (Maybe [HopDestination])

Optional list of hop destinations.

updateJobTemplate_queue :: Lens' UpdateJobTemplate (Maybe Text)

The new queue for the job template, if you are changing it.

updateJobTemplate_description :: Lens' UpdateJobTemplate (Maybe Text)

The new description for the job template, if you are changing it.

updateJobTemplate_name :: Lens' UpdateJobTemplate Text

The name of the job template you are modifying

Destructuring the Response

data UpdateJobTemplateResponse

See: newUpdateJobTemplateResponse smart constructor.

Constructors

UpdateJobTemplateResponse'

Fields

jobTemplate :: Maybe JobTemplate

A job template is a pre-made set of encoding instructions that you can use to quickly create a job.

httpStatus :: Int

The response's http status code.

newUpdateJobTemplateResponse

Arguments

:: Int
$sel:httpStatus:UpdateJobTemplateResponse'

-> UpdateJobTemplateResponse

Create a value of UpdateJobTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobTemplate:UpdateJobTemplateResponse', updateJobTemplateResponse_jobTemplate - A job template is a pre-made set of encoding instructions that you can use to quickly create a job.

$sel:httpStatus:UpdateJobTemplateResponse', updateJobTemplateResponse_httpStatus - The response's http status code.

Response Lenses

updateJobTemplateResponse_jobTemplate :: Lens' UpdateJobTemplateResponse (Maybe JobTemplate)

A job template is a pre-made set of encoding instructions that you can use to quickly create a job.

updateJobTemplateResponse_httpStatus :: Lens' UpdateJobTemplateResponse Int

The response's http status code.