Network.AWS.MediaConvert.UpdatePreset

Description

Modify one of your existing presets.

Creating a Request

data UpdatePreset

See: newUpdatePreset smart constructor.

Constructors

UpdatePreset'

Fields

category :: Maybe Text

The new category for the preset, if you are changing it.

description :: Maybe Text

The new description for the preset, if you are changing it.

settings :: Maybe PresetSettings

Settings for preset

name :: Text

The name of the preset you are modifying.

newUpdatePreset

Arguments

:: Text
$sel:name:UpdatePreset'

-> UpdatePreset

Create a value of UpdatePreset with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:category:UpdatePreset', updatePreset_category - The new category for the preset, if you are changing it.

$sel:description:UpdatePreset', updatePreset_description - The new description for the preset, if you are changing it.

$sel:settings:UpdatePreset', updatePreset_settings - Settings for preset

$sel:name:UpdatePreset', updatePreset_name - The name of the preset you are modifying.

Request Lenses

updatePreset_category :: Lens' UpdatePreset (Maybe Text)

The new category for the preset, if you are changing it.

updatePreset_description :: Lens' UpdatePreset (Maybe Text)

The new description for the preset, if you are changing it.

updatePreset_settings :: Lens' UpdatePreset (Maybe PresetSettings)

Settings for preset

updatePreset_name :: Lens' UpdatePreset Text

The name of the preset you are modifying.

Destructuring the Response

data UpdatePresetResponse

See: newUpdatePresetResponse smart constructor.

Constructors

UpdatePresetResponse'

Fields

preset :: Maybe Preset

A preset is a collection of preconfigured media conversion settings that you want MediaConvert to apply to the output during the conversion process.

httpStatus :: Int

The response's http status code.

newUpdatePresetResponse

Arguments

:: Int
$sel:httpStatus:UpdatePresetResponse'

-> UpdatePresetResponse

Create a value of UpdatePresetResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:preset:UpdatePresetResponse', updatePresetResponse_preset - A preset is a collection of preconfigured media conversion settings that you want MediaConvert to apply to the output during the conversion process.

$sel:httpStatus:UpdatePresetResponse', updatePresetResponse_httpStatus - The response's http status code.

Response Lenses

updatePresetResponse_preset :: Lens' UpdatePresetResponse (Maybe Preset)

A preset is a collection of preconfigured media conversion settings that you want MediaConvert to apply to the output during the conversion process.

updatePresetResponse_httpStatus :: Lens' UpdatePresetResponse Int

The response's http status code.