Network.AWS.MediaConvert.UpdateQueue

Description

Modify one of your existing queues.

Creating a Request

data UpdateQueue

See: newUpdateQueue smart constructor.

Constructors

UpdateQueue'

Fields

status :: Maybe QueueStatus

Pause or activate a queue by changing its status between ACTIVE and PAUSED. If you pause a queue, jobs in that queue won't begin. Jobs that are running when you pause the queue continue to run until they finish or result in an error.

reservationPlanSettings :: Maybe ReservationPlanSettings

The new details of your pricing plan for your reserved queue. When you set up a new pricing plan to replace an expired one, you enter into another 12-month commitment. When you add capacity to your queue by increasing the number of RTS, you extend the term of your commitment to 12 months from when you add capacity. After you make these commitments, you can't cancel them.

description :: Maybe Text

The new description for the queue, if you are changing it.

name :: Text

The name of the queue that you are modifying.

newUpdateQueue

Arguments

:: Text
$sel:name:UpdateQueue'

-> UpdateQueue

Create a value of UpdateQueue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:UpdateQueue', updateQueue_status - Pause or activate a queue by changing its status between ACTIVE and PAUSED. If you pause a queue, jobs in that queue won't begin. Jobs that are running when you pause the queue continue to run until they finish or result in an error.

$sel:reservationPlanSettings:UpdateQueue', updateQueue_reservationPlanSettings - The new details of your pricing plan for your reserved queue. When you set up a new pricing plan to replace an expired one, you enter into another 12-month commitment. When you add capacity to your queue by increasing the number of RTS, you extend the term of your commitment to 12 months from when you add capacity. After you make these commitments, you can't cancel them.

$sel:description:UpdateQueue', updateQueue_description - The new description for the queue, if you are changing it.

$sel:name:UpdateQueue', updateQueue_name - The name of the queue that you are modifying.

Request Lenses

updateQueue_status :: Lens' UpdateQueue (Maybe QueueStatus)

Pause or activate a queue by changing its status between ACTIVE and PAUSED. If you pause a queue, jobs in that queue won't begin. Jobs that are running when you pause the queue continue to run until they finish or result in an error.

updateQueue_reservationPlanSettings :: Lens' UpdateQueue (Maybe ReservationPlanSettings)

The new details of your pricing plan for your reserved queue. When you set up a new pricing plan to replace an expired one, you enter into another 12-month commitment. When you add capacity to your queue by increasing the number of RTS, you extend the term of your commitment to 12 months from when you add capacity. After you make these commitments, you can't cancel them.

updateQueue_description :: Lens' UpdateQueue (Maybe Text)

The new description for the queue, if you are changing it.

updateQueue_name :: Lens' UpdateQueue Text

The name of the queue that you are modifying.

Destructuring the Response

data UpdateQueueResponse

See: newUpdateQueueResponse smart constructor.

Constructors

UpdateQueueResponse'

Fields

queue :: Maybe Queue

You can use queues to manage the resources that are available to your AWS account for running multiple transcoding jobs at the same time. If you don't specify a queue, the service sends all jobs through the default queue. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/working-with-queues.html.

httpStatus :: Int

The response's http status code.

newUpdateQueueResponse

Arguments

:: Int
$sel:httpStatus:UpdateQueueResponse'

-> UpdateQueueResponse

Create a value of UpdateQueueResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:queue:UpdateQueueResponse', updateQueueResponse_queue - You can use queues to manage the resources that are available to your AWS account for running multiple transcoding jobs at the same time. If you don't specify a queue, the service sends all jobs through the default queue. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/working-with-queues.html.

$sel:httpStatus:UpdateQueueResponse', updateQueueResponse_httpStatus - The response's http status code.

Response Lenses

updateQueueResponse_queue :: Lens' UpdateQueueResponse (Maybe Queue)

You can use queues to manage the resources that are available to your AWS account for running multiple transcoding jobs at the same time. If you don't specify a queue, the service sends all jobs through the default queue. For more information, see https://docs.aws.amazon.com/mediaconvert/latest/ug/working-with-queues.html.

updateQueueResponse_httpStatus :: Lens' UpdateQueueResponse Int

The response's http status code.