Network.AWS.MediaLive.BatchDelete

Description

Starts delete of resources.

Creating a Request

data BatchDelete'

A request to delete resources

See: newBatchDelete' smart constructor.

Constructors

BatchDelete''

Fields

channelIds :: Maybe [Text]

List of channel IDs

inputIds :: Maybe [Text]

List of input IDs

multiplexIds :: Maybe [Text]

List of multiplex IDs

inputSecurityGroupIds :: Maybe [Text]

List of input security group IDs

newBatchDelete' :: BatchDelete'

Create a value of BatchDelete' with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelIds:BatchDelete'', batchDelete'_channelIds - List of channel IDs

$sel:inputIds:BatchDelete'', batchDelete'_inputIds - List of input IDs

$sel:multiplexIds:BatchDelete'', batchDelete'_multiplexIds - List of multiplex IDs

$sel:inputSecurityGroupIds:BatchDelete'', batchDelete'_inputSecurityGroupIds - List of input security group IDs

Request Lenses

batchDelete'_channelIds :: Lens' BatchDelete' (Maybe [Text])

List of channel IDs

batchDelete'_inputIds :: Lens' BatchDelete' (Maybe [Text])

List of input IDs

batchDelete'_multiplexIds :: Lens' BatchDelete' (Maybe [Text])

List of multiplex IDs

batchDelete'_inputSecurityGroupIds :: Lens' BatchDelete' (Maybe [Text])

List of input security group IDs

Destructuring the Response

data BatchDeleteResponse

Placeholder documentation for BatchDeleteResponse

See: newBatchDeleteResponse smart constructor.

Constructors

BatchDeleteResponse'

Fields

successful :: Maybe [BatchSuccessfulResultModel]

List of successful operations

failed :: Maybe [BatchFailedResultModel]

List of failed operations

httpStatus :: Int

The response's http status code.

newBatchDeleteResponse

Arguments

:: Int
$sel:httpStatus:BatchDeleteResponse'

-> BatchDeleteResponse

Create a value of BatchDeleteResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:successful:BatchDeleteResponse', batchDeleteResponse_successful - List of successful operations

$sel:failed:BatchDeleteResponse', batchDeleteResponse_failed - List of failed operations

$sel:httpStatus:BatchDeleteResponse', batchDeleteResponse_httpStatus - The response's http status code.

Response Lenses

batchDeleteResponse_successful :: Lens' BatchDeleteResponse (Maybe [BatchSuccessfulResultModel])

List of successful operations

batchDeleteResponse_failed :: Lens' BatchDeleteResponse (Maybe [BatchFailedResultModel])

List of failed operations

batchDeleteResponse_httpStatus :: Lens' BatchDeleteResponse Int

The response's http status code.