Network.AWS.MediaLive.CreateChannel

Description

Creates a new channel

Creating a Request

data CreateChannel'

A request to create a channel

See: newCreateChannel' smart constructor.

Constructors

CreateChannel''

Fields

roleArn :: Maybe Text

An optional Amazon Resource Name (ARN) of the role to assume when running the Channel.

encoderSettings :: Maybe EncoderSettings

inputSpecification :: Maybe InputSpecification

Specification of network and file inputs for this channel

channelClass :: Maybe ChannelClass

The class for this channel. STANDARD for a channel with two pipelines or SINGLE_PIPELINE for a channel with one pipeline.

logLevel :: Maybe LogLevel

The log level to write to CloudWatch Logs.

name :: Maybe Text

Name of channel.

destinations :: Maybe [OutputDestination]

inputAttachments :: Maybe [InputAttachment]

List of input attachments for channel.

reserved :: Maybe Text

Deprecated field that's only usable by whitelisted customers.

requestId :: Maybe Text

Unique request ID to be specified. This is needed to prevent retries from creating multiple resources.

tags :: Maybe (HashMap Text Text)

A collection of key-value pairs.

vpc :: Maybe VpcOutputSettings

Settings for VPC output

cdiInputSpecification :: Maybe CdiInputSpecification

Specification of CDI inputs for this channel

newCreateChannel' :: CreateChannel'

Create a value of CreateChannel' with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleArn:CreateChannel'', createChannel'_roleArn - An optional Amazon Resource Name (ARN) of the role to assume when running the Channel.

$sel:encoderSettings:CreateChannel'', createChannel'_encoderSettings - Undocumented member.

$sel:inputSpecification:CreateChannel'', createChannel'_inputSpecification - Specification of network and file inputs for this channel

$sel:channelClass:CreateChannel'', createChannel'_channelClass - The class for this channel. STANDARD for a channel with two pipelines or SINGLE_PIPELINE for a channel with one pipeline.

$sel:logLevel:CreateChannel'', createChannel'_logLevel - The log level to write to CloudWatch Logs.

$sel:name:CreateChannel'', createChannel'_name - Name of channel.

$sel:destinations:CreateChannel'', createChannel'_destinations - Undocumented member.

$sel:inputAttachments:CreateChannel'', createChannel'_inputAttachments - List of input attachments for channel.

$sel:reserved:CreateChannel'', createChannel'_reserved - Deprecated field that's only usable by whitelisted customers.

$sel:requestId:CreateChannel'', createChannel'_requestId - Unique request ID to be specified. This is needed to prevent retries from creating multiple resources.

$sel:tags:CreateChannel'', createChannel'_tags - A collection of key-value pairs.

$sel:vpc:CreateChannel'', createChannel'_vpc - Settings for VPC output

$sel:cdiInputSpecification:CreateChannel'', createChannel'_cdiInputSpecification - Specification of CDI inputs for this channel

Request Lenses

createChannel'_roleArn :: Lens' CreateChannel' (Maybe Text)

An optional Amazon Resource Name (ARN) of the role to assume when running the Channel.

createChannel'_encoderSettings :: Lens' CreateChannel' (Maybe EncoderSettings)

Undocumented member.

createChannel'_inputSpecification :: Lens' CreateChannel' (Maybe InputSpecification)

Specification of network and file inputs for this channel

createChannel'_channelClass :: Lens' CreateChannel' (Maybe ChannelClass)

The class for this channel. STANDARD for a channel with two pipelines or SINGLE_PIPELINE for a channel with one pipeline.

createChannel'_logLevel :: Lens' CreateChannel' (Maybe LogLevel)

The log level to write to CloudWatch Logs.

createChannel'_name :: Lens' CreateChannel' (Maybe Text)

Name of channel.

createChannel'_destinations :: Lens' CreateChannel' (Maybe [OutputDestination])

Undocumented member.

createChannel'_inputAttachments :: Lens' CreateChannel' (Maybe [InputAttachment])

List of input attachments for channel.

createChannel'_reserved :: Lens' CreateChannel' (Maybe Text)

Deprecated field that's only usable by whitelisted customers.

createChannel'_requestId :: Lens' CreateChannel' (Maybe Text)

Unique request ID to be specified. This is needed to prevent retries from creating multiple resources.

createChannel'_tags :: Lens' CreateChannel' (Maybe (HashMap Text Text))

A collection of key-value pairs.

createChannel'_vpc :: Lens' CreateChannel' (Maybe VpcOutputSettings)

Settings for VPC output

createChannel'_cdiInputSpecification :: Lens' CreateChannel' (Maybe CdiInputSpecification)

Specification of CDI inputs for this channel

Destructuring the Response

data CreateChannelResponse

Placeholder documentation for CreateChannelResponse

See: newCreateChannelResponse smart constructor.

Constructors

CreateChannelResponse'

Fields

channel :: Maybe Channel

httpStatus :: Int

The response's http status code.

newCreateChannelResponse

Arguments

:: Int
$sel:httpStatus:CreateChannelResponse'

-> CreateChannelResponse

Create a value of CreateChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channel:CreateChannelResponse', createChannelResponse_channel - Undocumented member.

$sel:httpStatus:CreateChannelResponse', createChannelResponse_httpStatus - The response's http status code.

Response Lenses

createChannelResponse_channel :: Lens' CreateChannelResponse (Maybe Channel)

Undocumented member.

createChannelResponse_httpStatus :: Lens' CreateChannelResponse Int

The response's http status code.