Network.AWS.MediaLive.CreateInput

Description

Create an input

Creating a Request

data CreateInput'

The name of the input

See: newCreateInput' smart constructor.

Constructors

CreateInput''

Fields

inputSecurityGroups :: Maybe [Text]

A list of security groups referenced by IDs to attach to the input.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the role this input assumes during and after creation.

sources :: Maybe [InputSourceRequest]

The source URLs for a PULL-type input. Every PULL type input needs exactly two source URLs for redundancy. Only specify sources for PULL type Inputs. Leave Destinations empty.

mediaConnectFlows :: Maybe [MediaConnectFlowRequest]

A list of the MediaConnect Flows that you want to use in this input. You can specify as few as one Flow and presently, as many as two. The only requirement is when you have more than one is that each Flow is in a separate Availability Zone as this ensures your EML input is redundant to AZ issues.

name :: Maybe Text

Name of the input.

destinations :: Maybe [InputDestinationRequest]

Destination settings for PUSH type inputs.

requestId :: Maybe Text

Unique identifier of the request to ensure the request is handled exactly once in case of retries.

tags :: Maybe (HashMap Text Text)

A collection of key-value pairs.

type' :: Maybe InputType

vpc :: Maybe InputVpcRequest

inputDevices :: Maybe [InputDeviceSettings]

Settings for the devices.

newCreateInput' :: CreateInput'

Create a value of CreateInput' with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputSecurityGroups:CreateInput'', createInput'_inputSecurityGroups - A list of security groups referenced by IDs to attach to the input.

$sel:roleArn:CreateInput'', createInput'_roleArn - The Amazon Resource Name (ARN) of the role this input assumes during and after creation.

$sel:sources:CreateInput'', createInput'_sources - The source URLs for a PULL-type input. Every PULL type input needs exactly two source URLs for redundancy. Only specify sources for PULL type Inputs. Leave Destinations empty.

$sel:mediaConnectFlows:CreateInput'', createInput'_mediaConnectFlows - A list of the MediaConnect Flows that you want to use in this input. You can specify as few as one Flow and presently, as many as two. The only requirement is when you have more than one is that each Flow is in a separate Availability Zone as this ensures your EML input is redundant to AZ issues.

$sel:name:CreateInput'', createInput'_name - Name of the input.

$sel:destinations:CreateInput'', createInput'_destinations - Destination settings for PUSH type inputs.

$sel:requestId:CreateInput'', createInput'_requestId - Unique identifier of the request to ensure the request is handled exactly once in case of retries.

$sel:tags:CreateInput'', createInput'_tags - A collection of key-value pairs.

$sel:type':CreateInput'', createInput'_type - Undocumented member.

$sel:vpc:CreateInput'', createInput'_vpc - Undocumented member.

$sel:inputDevices:CreateInput'', createInput'_inputDevices - Settings for the devices.

Request Lenses

createInput'_inputSecurityGroups :: Lens' CreateInput' (Maybe [Text])

A list of security groups referenced by IDs to attach to the input.

createInput'_roleArn :: Lens' CreateInput' (Maybe Text)

The Amazon Resource Name (ARN) of the role this input assumes during and after creation.

createInput'_sources :: Lens' CreateInput' (Maybe [InputSourceRequest])

The source URLs for a PULL-type input. Every PULL type input needs exactly two source URLs for redundancy. Only specify sources for PULL type Inputs. Leave Destinations empty.

createInput'_mediaConnectFlows :: Lens' CreateInput' (Maybe [MediaConnectFlowRequest])

A list of the MediaConnect Flows that you want to use in this input. You can specify as few as one Flow and presently, as many as two. The only requirement is when you have more than one is that each Flow is in a separate Availability Zone as this ensures your EML input is redundant to AZ issues.

createInput'_name :: Lens' CreateInput' (Maybe Text)

Name of the input.

createInput'_destinations :: Lens' CreateInput' (Maybe [InputDestinationRequest])

Destination settings for PUSH type inputs.

createInput'_requestId :: Lens' CreateInput' (Maybe Text)

Unique identifier of the request to ensure the request is handled exactly once in case of retries.

createInput'_tags :: Lens' CreateInput' (Maybe (HashMap Text Text))

A collection of key-value pairs.

createInput'_type :: Lens' CreateInput' (Maybe InputType)

Undocumented member.

createInput'_vpc :: Lens' CreateInput' (Maybe InputVpcRequest)

Undocumented member.

createInput'_inputDevices :: Lens' CreateInput' (Maybe [InputDeviceSettings])

Settings for the devices.

Destructuring the Response

data CreateInputResponse

Placeholder documentation for CreateInputResponse

See: newCreateInputResponse smart constructor.

Constructors

CreateInputResponse'

Fields

input :: Maybe Input

httpStatus :: Int

The response's http status code.

newCreateInputResponse

Arguments

:: Int
$sel:httpStatus:CreateInputResponse'

-> CreateInputResponse

Create a value of CreateInputResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:input:CreateInputResponse', createInputResponse_input - Undocumented member.

$sel:httpStatus:CreateInputResponse', createInputResponse_httpStatus - The response's http status code.

Response Lenses

createInputResponse_input :: Lens' CreateInputResponse (Maybe Input)

Undocumented member.

createInputResponse_httpStatus :: Lens' CreateInputResponse Int

The response's http status code.