Network.AWS.MediaLive.CreateMultiplex

Description

Create a new multiplex.

Creating a Request

data CreateMultiplex'

A request to create a multiplex.

See: newCreateMultiplex' smart constructor.

Constructors

CreateMultiplex''

Fields

tags :: Maybe (HashMap Text Text)

A collection of key-value pairs.

requestId :: Text

Unique request ID. This prevents retries from creating multiple resources.

multiplexSettings :: MultiplexSettings

Configuration for a multiplex event.

availabilityZones :: [Text]

A list of availability zones for the multiplex. You must specify exactly two.

name :: Text

Name of multiplex.

newCreateMultiplex'

Arguments

:: Text
$sel:requestId:CreateMultiplex''

-> MultiplexSettings
$sel:multiplexSettings:CreateMultiplex''

-> Text
$sel:name:CreateMultiplex''

-> CreateMultiplex'

Create a value of CreateMultiplex' with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateMultiplex'', createMultiplex'_tags - A collection of key-value pairs.

$sel:requestId:CreateMultiplex'', createMultiplex'_requestId - Unique request ID. This prevents retries from creating multiple resources.

$sel:multiplexSettings:CreateMultiplex'', createMultiplex'_multiplexSettings - Configuration for a multiplex event.

$sel:availabilityZones:CreateMultiplex'', createMultiplex'_availabilityZones - A list of availability zones for the multiplex. You must specify exactly two.

$sel:name:CreateMultiplex'', createMultiplex'_name - Name of multiplex.

Request Lenses

createMultiplex'_tags :: Lens' CreateMultiplex' (Maybe (HashMap Text Text))

A collection of key-value pairs.

createMultiplex'_requestId :: Lens' CreateMultiplex' Text

Unique request ID. This prevents retries from creating multiple resources.

createMultiplex'_multiplexSettings :: Lens' CreateMultiplex' MultiplexSettings

Configuration for a multiplex event.

createMultiplex'_availabilityZones :: Lens' CreateMultiplex' [Text]

A list of availability zones for the multiplex. You must specify exactly two.

createMultiplex'_name :: Lens' CreateMultiplex' Text

Name of multiplex.

Destructuring the Response

data CreateMultiplexResponse

Placeholder documentation for CreateMultiplexResponse

See: newCreateMultiplexResponse smart constructor.

Constructors

CreateMultiplexResponse'

Fields

multiplex :: Maybe Multiplex

The newly created multiplex.

httpStatus :: Int

The response's http status code.

newCreateMultiplexResponse

Arguments

:: Int
$sel:httpStatus:CreateMultiplexResponse'

-> CreateMultiplexResponse

Create a value of CreateMultiplexResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:multiplex:CreateMultiplexResponse', createMultiplexResponse_multiplex - The newly created multiplex.

$sel:httpStatus:CreateMultiplexResponse', createMultiplexResponse_httpStatus - The response's http status code.

Response Lenses

createMultiplexResponse_multiplex :: Lens' CreateMultiplexResponse (Maybe Multiplex)

The newly created multiplex.

createMultiplexResponse_httpStatus :: Lens' CreateMultiplexResponse Int

The response's http status code.