Network.AWS.MediaLive.CreateMultiplexProgram

Description

Create a new program in the multiplex.

Creating a Request

data CreateMultiplexProgram'

A request to create a program in a multiplex.

See: newCreateMultiplexProgram' smart constructor.

Constructors

CreateMultiplexProgram''

Fields

multiplexId :: Text

ID of the multiplex where the program is to be created.

requestId :: Text

Unique request ID. This prevents retries from creating multiple resources.

multiplexProgramSettings :: MultiplexProgramSettings

The settings for this multiplex program.

programName :: Text

Name of multiplex program.

newCreateMultiplexProgram'

Arguments

:: Text
$sel:multiplexId:CreateMultiplexProgram''

-> Text
$sel:requestId:CreateMultiplexProgram''

-> MultiplexProgramSettings
$sel:multiplexProgramSettings:CreateMultiplexProgram''

-> Text
$sel:programName:CreateMultiplexProgram''

-> CreateMultiplexProgram'

Create a value of CreateMultiplexProgram' with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:multiplexId:CreateMultiplexProgram'', createMultiplexProgram'_multiplexId - ID of the multiplex where the program is to be created.

$sel:requestId:CreateMultiplexProgram'', createMultiplexProgram'_requestId - Unique request ID. This prevents retries from creating multiple resources.

$sel:multiplexProgramSettings:CreateMultiplexProgram'', createMultiplexProgram'_multiplexProgramSettings - The settings for this multiplex program.

$sel:programName:CreateMultiplexProgram'', createMultiplexProgram'_programName - Name of multiplex program.

Request Lenses

createMultiplexProgram'_multiplexId :: Lens' CreateMultiplexProgram' Text

ID of the multiplex where the program is to be created.

createMultiplexProgram'_requestId :: Lens' CreateMultiplexProgram' Text

Unique request ID. This prevents retries from creating multiple resources.

createMultiplexProgram'_multiplexProgramSettings :: Lens' CreateMultiplexProgram' MultiplexProgramSettings

The settings for this multiplex program.

createMultiplexProgram'_programName :: Lens' CreateMultiplexProgram' Text

Name of multiplex program.

Destructuring the Response

data CreateMultiplexProgramResponse

Placeholder documentation for CreateMultiplexProgramResponse

See: newCreateMultiplexProgramResponse smart constructor.

Constructors

CreateMultiplexProgramResponse'

Fields

multiplexProgram :: Maybe MultiplexProgram

The newly created multiplex program.

httpStatus :: Int

The response's http status code.

newCreateMultiplexProgramResponse

Arguments

:: Int
$sel:httpStatus:CreateMultiplexProgramResponse'

-> CreateMultiplexProgramResponse

Create a value of CreateMultiplexProgramResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:multiplexProgram:CreateMultiplexProgramResponse', createMultiplexProgramResponse_multiplexProgram - The newly created multiplex program.

$sel:httpStatus:CreateMultiplexProgramResponse', createMultiplexProgramResponse_httpStatus - The response's http status code.

Response Lenses

createMultiplexProgramResponse_multiplexProgram :: Lens' CreateMultiplexProgramResponse (Maybe MultiplexProgram)

The newly created multiplex program.

createMultiplexProgramResponse_httpStatus :: Lens' CreateMultiplexProgramResponse Int

The response's http status code.