Network.AWS.MediaLive.CreatePartnerInput

Description

Create a partner input

Creating a Request

data CreatePartnerInput'

A request to create a partner input

See: newCreatePartnerInput' smart constructor.

Constructors

CreatePartnerInput''

Fields

requestId :: Maybe Text

Unique identifier of the request to ensure the request is handled exactly once in case of retries.

tags :: Maybe (HashMap Text Text)

A collection of key-value pairs.

inputId :: Text

Unique ID of the input.

newCreatePartnerInput'

Arguments

:: Text
$sel:inputId:CreatePartnerInput''

-> CreatePartnerInput'

Create a value of CreatePartnerInput' with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:CreatePartnerInput'', createPartnerInput'_requestId - Unique identifier of the request to ensure the request is handled exactly once in case of retries.

$sel:tags:CreatePartnerInput'', createPartnerInput'_tags - A collection of key-value pairs.

$sel:inputId:CreatePartnerInput'', createPartnerInput'_inputId - Unique ID of the input.

Request Lenses

createPartnerInput'_requestId :: Lens' CreatePartnerInput' (Maybe Text)

Unique identifier of the request to ensure the request is handled exactly once in case of retries.

createPartnerInput'_tags :: Lens' CreatePartnerInput' (Maybe (HashMap Text Text))

A collection of key-value pairs.

createPartnerInput'_inputId :: Lens' CreatePartnerInput' Text

Unique ID of the input.

Destructuring the Response

data CreatePartnerInputResponse

Placeholder documentation for CreatePartnerInputResponse

See: newCreatePartnerInputResponse smart constructor.

Constructors

CreatePartnerInputResponse'

Fields

input :: Maybe Input

httpStatus :: Int

The response's http status code.

newCreatePartnerInputResponse

Arguments

:: Int
$sel:httpStatus:CreatePartnerInputResponse'

-> CreatePartnerInputResponse

Create a value of CreatePartnerInputResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:input:CreatePartnerInputResponse', createPartnerInputResponse_input - Undocumented member.

$sel:httpStatus:CreatePartnerInputResponse', createPartnerInputResponse_httpStatus - The response's http status code.

Response Lenses

createPartnerInputResponse_input :: Lens' CreatePartnerInputResponse (Maybe Input)

Undocumented member.

createPartnerInputResponse_httpStatus :: Lens' CreatePartnerInputResponse Int

The response's http status code.