Network.AWS.MediaLive.DeleteChannel

Description

Starts deletion of channel. The associated outputs are also deleted.

Creating a Request

data DeleteChannel

Placeholder documentation for DeleteChannelRequest

See: newDeleteChannel smart constructor.

Constructors

DeleteChannel'

Fields

channelId :: Text

Unique ID of the channel.

newDeleteChannel

Arguments

:: Text
$sel:channelId:DeleteChannel'

-> DeleteChannel

Create a value of DeleteChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelId:DeleteChannel', deleteChannel_channelId - Unique ID of the channel.

Request Lenses

deleteChannel_channelId :: Lens' DeleteChannel Text

Unique ID of the channel.

Destructuring the Response

data DeleteChannelResponse

Placeholder documentation for DeleteChannelResponse

See: newDeleteChannelResponse smart constructor.

Constructors

DeleteChannelResponse'

Fields

state :: Maybe ChannelState

logLevel :: Maybe LogLevel

The log level being written to CloudWatch Logs.

arn :: Maybe Text

The unique arn of the channel.

pipelinesRunningCount :: Maybe Int

The number of currently healthy pipelines.

pipelineDetails :: Maybe [PipelineDetail]

Runtime details for the pipelines of a running channel.

inputSpecification :: Maybe InputSpecification

Specification of network and file inputs for this channel

inputAttachments :: Maybe [InputAttachment]

List of input attachments for channel.

destinations :: Maybe [OutputDestination]

A list of destinations of the channel. For UDP outputs, there is one destination per output. For other types (HLS, for example), there is one destination per packager.

name :: Maybe Text

The name of the channel. (user-mutable)

cdiInputSpecification :: Maybe CdiInputSpecification

Specification of CDI inputs for this channel

id :: Maybe Text

The unique id of the channel.

channelClass :: Maybe ChannelClass

The class for this channel. STANDARD for a channel with two pipelines or SINGLE_PIPELINE for a channel with one pipeline.

vpc :: Maybe VpcOutputSettingsDescription

Settings for VPC output

egressEndpoints :: Maybe [ChannelEgressEndpoint]

The endpoints where outgoing connections initiate from

tags :: Maybe (HashMap Text Text)

A collection of key-value pairs.

encoderSettings :: Maybe EncoderSettings

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the role assumed when running the Channel.

httpStatus :: Int

The response's http status code.

newDeleteChannelResponse

Arguments

:: Int
$sel:httpStatus:DeleteChannelResponse'

-> DeleteChannelResponse

Create a value of DeleteChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:DeleteChannelResponse', deleteChannelResponse_state - Undocumented member.

$sel:logLevel:DeleteChannelResponse', deleteChannelResponse_logLevel - The log level being written to CloudWatch Logs.

$sel:arn:DeleteChannelResponse', deleteChannelResponse_arn - The unique arn of the channel.

$sel:pipelinesRunningCount:DeleteChannelResponse', deleteChannelResponse_pipelinesRunningCount - The number of currently healthy pipelines.

$sel:pipelineDetails:DeleteChannelResponse', deleteChannelResponse_pipelineDetails - Runtime details for the pipelines of a running channel.

$sel:inputSpecification:DeleteChannelResponse', deleteChannelResponse_inputSpecification - Specification of network and file inputs for this channel

$sel:inputAttachments:DeleteChannelResponse', deleteChannelResponse_inputAttachments - List of input attachments for channel.

$sel:destinations:DeleteChannelResponse', deleteChannelResponse_destinations - A list of destinations of the channel. For UDP outputs, there is one destination per output. For other types (HLS, for example), there is one destination per packager.

$sel:name:DeleteChannelResponse', deleteChannelResponse_name - The name of the channel. (user-mutable)

$sel:cdiInputSpecification:DeleteChannelResponse', deleteChannelResponse_cdiInputSpecification - Specification of CDI inputs for this channel

$sel:id:DeleteChannelResponse', deleteChannelResponse_id - The unique id of the channel.

$sel:channelClass:DeleteChannelResponse', deleteChannelResponse_channelClass - The class for this channel. STANDARD for a channel with two pipelines or SINGLE_PIPELINE for a channel with one pipeline.

$sel:vpc:DeleteChannelResponse', deleteChannelResponse_vpc - Settings for VPC output

$sel:egressEndpoints:DeleteChannelResponse', deleteChannelResponse_egressEndpoints - The endpoints where outgoing connections initiate from

$sel:tags:DeleteChannelResponse', deleteChannelResponse_tags - A collection of key-value pairs.

$sel:encoderSettings:DeleteChannelResponse', deleteChannelResponse_encoderSettings - Undocumented member.

$sel:roleArn:DeleteChannelResponse', deleteChannelResponse_roleArn - The Amazon Resource Name (ARN) of the role assumed when running the Channel.

$sel:httpStatus:DeleteChannelResponse', deleteChannelResponse_httpStatus - The response's http status code.

Response Lenses

deleteChannelResponse_state :: Lens' DeleteChannelResponse (Maybe ChannelState)

Undocumented member.

deleteChannelResponse_logLevel :: Lens' DeleteChannelResponse (Maybe LogLevel)

The log level being written to CloudWatch Logs.

deleteChannelResponse_arn :: Lens' DeleteChannelResponse (Maybe Text)

The unique arn of the channel.

deleteChannelResponse_pipelinesRunningCount :: Lens' DeleteChannelResponse (Maybe Int)

The number of currently healthy pipelines.

deleteChannelResponse_pipelineDetails :: Lens' DeleteChannelResponse (Maybe [PipelineDetail])

Runtime details for the pipelines of a running channel.

deleteChannelResponse_inputSpecification :: Lens' DeleteChannelResponse (Maybe InputSpecification)

Specification of network and file inputs for this channel

deleteChannelResponse_inputAttachments :: Lens' DeleteChannelResponse (Maybe [InputAttachment])

List of input attachments for channel.

deleteChannelResponse_destinations :: Lens' DeleteChannelResponse (Maybe [OutputDestination])

A list of destinations of the channel. For UDP outputs, there is one destination per output. For other types (HLS, for example), there is one destination per packager.

deleteChannelResponse_name :: Lens' DeleteChannelResponse (Maybe Text)

The name of the channel. (user-mutable)

deleteChannelResponse_cdiInputSpecification :: Lens' DeleteChannelResponse (Maybe CdiInputSpecification)

Specification of CDI inputs for this channel

deleteChannelResponse_id :: Lens' DeleteChannelResponse (Maybe Text)

The unique id of the channel.

deleteChannelResponse_channelClass :: Lens' DeleteChannelResponse (Maybe ChannelClass)

The class for this channel. STANDARD for a channel with two pipelines or SINGLE_PIPELINE for a channel with one pipeline.

deleteChannelResponse_vpc :: Lens' DeleteChannelResponse (Maybe VpcOutputSettingsDescription)

Settings for VPC output

deleteChannelResponse_egressEndpoints :: Lens' DeleteChannelResponse (Maybe [ChannelEgressEndpoint])

The endpoints where outgoing connections initiate from

deleteChannelResponse_tags :: Lens' DeleteChannelResponse (Maybe (HashMap Text Text))

A collection of key-value pairs.

deleteChannelResponse_encoderSettings :: Lens' DeleteChannelResponse (Maybe EncoderSettings)

Undocumented member.

deleteChannelResponse_roleArn :: Lens' DeleteChannelResponse (Maybe Text)

The Amazon Resource Name (ARN) of the role assumed when running the Channel.

deleteChannelResponse_httpStatus :: Lens' DeleteChannelResponse Int

The response's http status code.