Network.AWS.MediaLive.DeleteReservation

Description

Delete an expired reservation.

Creating a Request

data DeleteReservation

Placeholder documentation for DeleteReservationRequest

See: newDeleteReservation smart constructor.

Constructors

DeleteReservation'

Fields

reservationId :: Text

Unique reservation ID, e.g. '1234567'

newDeleteReservation

Arguments

:: Text
$sel:reservationId:DeleteReservation'

-> DeleteReservation

Create a value of DeleteReservation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reservationId:DeleteReservation', deleteReservation_reservationId - Unique reservation ID, e.g. '1234567'

Request Lenses

deleteReservation_reservationId :: Lens' DeleteReservation Text

Unique reservation ID, e.g. '1234567'

Destructuring the Response

data DeleteReservationResponse

Placeholder documentation for DeleteReservationResponse

See: newDeleteReservationResponse smart constructor.

Constructors

DeleteReservationResponse'

Fields

end :: Maybe Text

Reservation UTC end date and time in ISO-8601 format, e.g. '2019-03-01T00:00:00'

duration :: Maybe Int

Lease duration, e.g. '12'

durationUnits :: Maybe OfferingDurationUnits

Units for duration, e.g. 'MONTHS'

arn :: Maybe Text

Unique reservation ARN, e.g. 'arn:aws:medialive:us-west-2:123456789012:reservation:1234567'

offeringId :: Maybe Text

Unique offering ID, e.g. '87654321'

currencyCode :: Maybe Text

Currency code for usagePrice and fixedPrice in ISO-4217 format, e.g. 'USD'

name :: Maybe Text

User specified reservation name

resourceSpecification :: Maybe ReservationResourceSpecification

Resource configuration details

state :: Maybe ReservationState

Current state of reservation, e.g. 'ACTIVE'

tags :: Maybe (HashMap Text Text)

A collection of key-value pairs

offeringDescription :: Maybe Text

Offering description, e.g. 'HD AVC output at 10-20 Mbps, 30 fps, and standard VQ in US West (Oregon)'

fixedPrice :: Maybe Double

One-time charge for each reserved resource, e.g. '0.0' for a NO_UPFRONT offering

count :: Maybe Int

Number of reserved resources

usagePrice :: Maybe Double

Recurring usage charge for each reserved resource, e.g. '157.0'

offeringType :: Maybe OfferingType

Offering type, e.g. 'NO_UPFRONT'

region :: Maybe Text

AWS region, e.g. 'us-west-2'

start :: Maybe Text

Reservation UTC start date and time in ISO-8601 format, e.g. '2018-03-01T00:00:00'

reservationId :: Maybe Text

Unique reservation ID, e.g. '1234567'

httpStatus :: Int

The response's http status code.

newDeleteReservationResponse

Arguments

:: Int
$sel:httpStatus:DeleteReservationResponse'

-> DeleteReservationResponse

Create a value of DeleteReservationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:end:DeleteReservationResponse', deleteReservationResponse_end - Reservation UTC end date and time in ISO-8601 format, e.g. '2019-03-01T00:00:00'

$sel:duration:DeleteReservationResponse', deleteReservationResponse_duration - Lease duration, e.g. '12'

$sel:durationUnits:DeleteReservationResponse', deleteReservationResponse_durationUnits - Units for duration, e.g. 'MONTHS'

$sel:arn:DeleteReservationResponse', deleteReservationResponse_arn - Unique reservation ARN, e.g. 'arn:aws:medialive:us-west-2:123456789012:reservation:1234567'

$sel:offeringId:DeleteReservationResponse', deleteReservationResponse_offeringId - Unique offering ID, e.g. '87654321'

$sel:currencyCode:DeleteReservationResponse', deleteReservationResponse_currencyCode - Currency code for usagePrice and fixedPrice in ISO-4217 format, e.g. 'USD'

$sel:name:DeleteReservationResponse', deleteReservationResponse_name - User specified reservation name

$sel:resourceSpecification:DeleteReservationResponse', deleteReservationResponse_resourceSpecification - Resource configuration details

$sel:state:DeleteReservationResponse', deleteReservationResponse_state - Current state of reservation, e.g. 'ACTIVE'

$sel:tags:DeleteReservationResponse', deleteReservationResponse_tags - A collection of key-value pairs

$sel:offeringDescription:DeleteReservationResponse', deleteReservationResponse_offeringDescription - Offering description, e.g. 'HD AVC output at 10-20 Mbps, 30 fps, and standard VQ in US West (Oregon)'

$sel:fixedPrice:DeleteReservationResponse', deleteReservationResponse_fixedPrice - One-time charge for each reserved resource, e.g. '0.0' for a NO_UPFRONT offering

$sel:count:DeleteReservationResponse', deleteReservationResponse_count - Number of reserved resources

$sel:usagePrice:DeleteReservationResponse', deleteReservationResponse_usagePrice - Recurring usage charge for each reserved resource, e.g. '157.0'

$sel:offeringType:DeleteReservationResponse', deleteReservationResponse_offeringType - Offering type, e.g. 'NO_UPFRONT'

$sel:region:DeleteReservationResponse', deleteReservationResponse_region - AWS region, e.g. 'us-west-2'

$sel:start:DeleteReservationResponse', deleteReservationResponse_start - Reservation UTC start date and time in ISO-8601 format, e.g. '2018-03-01T00:00:00'

$sel:reservationId:DeleteReservation', deleteReservationResponse_reservationId - Unique reservation ID, e.g. '1234567'

$sel:httpStatus:DeleteReservationResponse', deleteReservationResponse_httpStatus - The response's http status code.

Response Lenses

deleteReservationResponse_end :: Lens' DeleteReservationResponse (Maybe Text)

Reservation UTC end date and time in ISO-8601 format, e.g. '2019-03-01T00:00:00'

deleteReservationResponse_duration :: Lens' DeleteReservationResponse (Maybe Int)

Lease duration, e.g. '12'

deleteReservationResponse_durationUnits :: Lens' DeleteReservationResponse (Maybe OfferingDurationUnits)

Units for duration, e.g. 'MONTHS'

deleteReservationResponse_arn :: Lens' DeleteReservationResponse (Maybe Text)

Unique reservation ARN, e.g. 'arn:aws:medialive:us-west-2:123456789012:reservation:1234567'

deleteReservationResponse_offeringId :: Lens' DeleteReservationResponse (Maybe Text)

Unique offering ID, e.g. '87654321'

deleteReservationResponse_currencyCode :: Lens' DeleteReservationResponse (Maybe Text)

Currency code for usagePrice and fixedPrice in ISO-4217 format, e.g. 'USD'

deleteReservationResponse_name :: Lens' DeleteReservationResponse (Maybe Text)

User specified reservation name

deleteReservationResponse_resourceSpecification :: Lens' DeleteReservationResponse (Maybe ReservationResourceSpecification)

Resource configuration details

deleteReservationResponse_state :: Lens' DeleteReservationResponse (Maybe ReservationState)

Current state of reservation, e.g. 'ACTIVE'

deleteReservationResponse_tags :: Lens' DeleteReservationResponse (Maybe (HashMap Text Text))

A collection of key-value pairs

deleteReservationResponse_offeringDescription :: Lens' DeleteReservationResponse (Maybe Text)

Offering description, e.g. 'HD AVC output at 10-20 Mbps, 30 fps, and standard VQ in US West (Oregon)'

deleteReservationResponse_fixedPrice :: Lens' DeleteReservationResponse (Maybe Double)

One-time charge for each reserved resource, e.g. '0.0' for a NO_UPFRONT offering

deleteReservationResponse_count :: Lens' DeleteReservationResponse (Maybe Int)

Number of reserved resources

deleteReservationResponse_usagePrice :: Lens' DeleteReservationResponse (Maybe Double)

Recurring usage charge for each reserved resource, e.g. '157.0'

deleteReservationResponse_offeringType :: Lens' DeleteReservationResponse (Maybe OfferingType)

Offering type, e.g. 'NO_UPFRONT'

deleteReservationResponse_region :: Lens' DeleteReservationResponse (Maybe Text)

AWS region, e.g. 'us-west-2'

deleteReservationResponse_start :: Lens' DeleteReservationResponse (Maybe Text)

Reservation UTC start date and time in ISO-8601 format, e.g. '2018-03-01T00:00:00'

deleteReservationResponse_reservationId :: Lens' DeleteReservationResponse (Maybe Text)

Unique reservation ID, e.g. '1234567'

deleteReservationResponse_httpStatus :: Lens' DeleteReservationResponse Int

The response's http status code.