Network.AWS.MediaLive.DeleteSchedule

Description

Delete all schedule actions on a channel.

Creating a Request

data DeleteSchedule

Placeholder documentation for DeleteScheduleRequest

See: newDeleteSchedule smart constructor.

Constructors

DeleteSchedule'

Fields

channelId :: Text

Id of the channel whose schedule is being deleted.

newDeleteSchedule

Arguments

:: Text
$sel:channelId:DeleteSchedule'

-> DeleteSchedule

Create a value of DeleteSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelId:DeleteSchedule', deleteSchedule_channelId - Id of the channel whose schedule is being deleted.

Request Lenses

deleteSchedule_channelId :: Lens' DeleteSchedule Text

Id of the channel whose schedule is being deleted.

Destructuring the Response

data DeleteScheduleResponse

Placeholder documentation for DeleteScheduleResponse

See: newDeleteScheduleResponse smart constructor.

Constructors

DeleteScheduleResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteScheduleResponse

Arguments

:: Int
$sel:httpStatus:DeleteScheduleResponse'

-> DeleteScheduleResponse

Create a value of DeleteScheduleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteScheduleResponse', deleteScheduleResponse_httpStatus - The response's http status code.

Response Lenses

deleteScheduleResponse_httpStatus :: Lens' DeleteScheduleResponse Int

The response's http status code.