Network.AWS.MediaLive.DescribeInput

Description

Produces details about an input

Creating a Request

data DescribeInput

Placeholder documentation for DescribeInputRequest

See: newDescribeInput smart constructor.

Constructors

DescribeInput'

Fields

inputId :: Text

Unique ID of the input

newDescribeInput

Arguments

:: Text
$sel:inputId:DescribeInput'

-> DescribeInput

Create a value of DescribeInput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputId:DescribeInput', describeInput_inputId - Unique ID of the input

Request Lenses

describeInput_inputId :: Lens' DescribeInput Text

Unique ID of the input

Destructuring the Response

data DescribeInputResponse

Placeholder documentation for DescribeInputResponse

See: newDescribeInputResponse smart constructor.

Constructors

DescribeInputResponse'

Fields

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the role this input assumes during and after creation.

inputPartnerIds :: Maybe [Text]

A list of IDs for all Inputs which are partners of this one.

sources :: Maybe [InputSource]

A list of the sources of the input (PULL-type).

mediaConnectFlows :: Maybe [MediaConnectFlow]

A list of MediaConnect Flows for this input.

inputSourceType :: Maybe InputSourceType

Certain pull input sources can be dynamic, meaning that they can have their URL's dynamically changes during input switch actions. Presently, this functionality only works with MP4_FILE inputs.

arn :: Maybe Text

The Unique ARN of the input (generated, immutable).

id :: Maybe Text

The generated ID of the input (unique for user account, immutable).

securityGroups :: Maybe [Text]

A list of IDs for all the Input Security Groups attached to the input.

name :: Maybe Text

The user-assigned name (This is a mutable value).

destinations :: Maybe [InputDestination]

A list of the destinations of the input (PUSH-type).

state :: Maybe InputState

tags :: Maybe (HashMap Text Text)

A collection of key-value pairs.

inputClass :: Maybe InputClass

STANDARD - MediaLive expects two sources to be connected to this input. If the channel is also STANDARD, both sources will be ingested. If the channel is SINGLE_PIPELINE, only the first source will be ingested; the second source will always be ignored, even if the first source fails. SINGLE_PIPELINE - You can connect only one source to this input. If the ChannelClass is also SINGLE_PIPELINE, this value is valid. If the ChannelClass is STANDARD, this value is not valid because the channel requires two sources in the input.

type' :: Maybe InputType

inputDevices :: Maybe [InputDeviceSettings]

Settings for the input devices.

attachedChannels :: Maybe [Text]

A list of channel IDs that that input is attached to (currently an input can only be attached to one channel).

httpStatus :: Int

The response's http status code.

newDescribeInputResponse

Arguments

:: Int
$sel:httpStatus:DescribeInputResponse'

-> DescribeInputResponse

Create a value of DescribeInputResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleArn:DescribeInputResponse', describeInputResponse_roleArn - The Amazon Resource Name (ARN) of the role this input assumes during and after creation.

$sel:inputPartnerIds:DescribeInputResponse', describeInputResponse_inputPartnerIds - A list of IDs for all Inputs which are partners of this one.

$sel:sources:DescribeInputResponse', describeInputResponse_sources - A list of the sources of the input (PULL-type).

$sel:mediaConnectFlows:DescribeInputResponse', describeInputResponse_mediaConnectFlows - A list of MediaConnect Flows for this input.

$sel:inputSourceType:DescribeInputResponse', describeInputResponse_inputSourceType - Certain pull input sources can be dynamic, meaning that they can have their URL's dynamically changes during input switch actions. Presently, this functionality only works with MP4_FILE inputs.

$sel:arn:DescribeInputResponse', describeInputResponse_arn - The Unique ARN of the input (generated, immutable).

$sel:id:DescribeInputResponse', describeInputResponse_id - The generated ID of the input (unique for user account, immutable).

$sel:securityGroups:DescribeInputResponse', describeInputResponse_securityGroups - A list of IDs for all the Input Security Groups attached to the input.

$sel:name:DescribeInputResponse', describeInputResponse_name - The user-assigned name (This is a mutable value).

$sel:destinations:DescribeInputResponse', describeInputResponse_destinations - A list of the destinations of the input (PUSH-type).

$sel:state:DescribeInputResponse', describeInputResponse_state - Undocumented member.

$sel:tags:DescribeInputResponse', describeInputResponse_tags - A collection of key-value pairs.

$sel:inputClass:DescribeInputResponse', describeInputResponse_inputClass - STANDARD - MediaLive expects two sources to be connected to this input. If the channel is also STANDARD, both sources will be ingested. If the channel is SINGLE_PIPELINE, only the first source will be ingested; the second source will always be ignored, even if the first source fails. SINGLE_PIPELINE - You can connect only one source to this input. If the ChannelClass is also SINGLE_PIPELINE, this value is valid. If the ChannelClass is STANDARD, this value is not valid because the channel requires two sources in the input.

$sel:type':DescribeInputResponse', describeInputResponse_type - Undocumented member.

$sel:inputDevices:DescribeInputResponse', describeInputResponse_inputDevices - Settings for the input devices.

$sel:attachedChannels:DescribeInputResponse', describeInputResponse_attachedChannels - A list of channel IDs that that input is attached to (currently an input can only be attached to one channel).

$sel:httpStatus:DescribeInputResponse', describeInputResponse_httpStatus - The response's http status code.

Response Lenses

describeInputResponse_roleArn :: Lens' DescribeInputResponse (Maybe Text)

The Amazon Resource Name (ARN) of the role this input assumes during and after creation.

describeInputResponse_inputPartnerIds :: Lens' DescribeInputResponse (Maybe [Text])

A list of IDs for all Inputs which are partners of this one.

describeInputResponse_sources :: Lens' DescribeInputResponse (Maybe [InputSource])

A list of the sources of the input (PULL-type).

describeInputResponse_mediaConnectFlows :: Lens' DescribeInputResponse (Maybe [MediaConnectFlow])

A list of MediaConnect Flows for this input.

describeInputResponse_inputSourceType :: Lens' DescribeInputResponse (Maybe InputSourceType)

Certain pull input sources can be dynamic, meaning that they can have their URL's dynamically changes during input switch actions. Presently, this functionality only works with MP4_FILE inputs.

describeInputResponse_arn :: Lens' DescribeInputResponse (Maybe Text)

The Unique ARN of the input (generated, immutable).

describeInputResponse_id :: Lens' DescribeInputResponse (Maybe Text)

The generated ID of the input (unique for user account, immutable).

describeInputResponse_securityGroups :: Lens' DescribeInputResponse (Maybe [Text])

A list of IDs for all the Input Security Groups attached to the input.

describeInputResponse_name :: Lens' DescribeInputResponse (Maybe Text)

The user-assigned name (This is a mutable value).

describeInputResponse_destinations :: Lens' DescribeInputResponse (Maybe [InputDestination])

A list of the destinations of the input (PUSH-type).

describeInputResponse_state :: Lens' DescribeInputResponse (Maybe InputState)

Undocumented member.

describeInputResponse_tags :: Lens' DescribeInputResponse (Maybe (HashMap Text Text))

A collection of key-value pairs.

describeInputResponse_inputClass :: Lens' DescribeInputResponse (Maybe InputClass)

STANDARD - MediaLive expects two sources to be connected to this input. If the channel is also STANDARD, both sources will be ingested. If the channel is SINGLE_PIPELINE, only the first source will be ingested; the second source will always be ignored, even if the first source fails. SINGLE_PIPELINE - You can connect only one source to this input. If the ChannelClass is also SINGLE_PIPELINE, this value is valid. If the ChannelClass is STANDARD, this value is not valid because the channel requires two sources in the input.

describeInputResponse_type :: Lens' DescribeInputResponse (Maybe InputType)

Undocumented member.

describeInputResponse_inputDevices :: Lens' DescribeInputResponse (Maybe [InputDeviceSettings])

Settings for the input devices.

describeInputResponse_attachedChannels :: Lens' DescribeInputResponse (Maybe [Text])

A list of channel IDs that that input is attached to (currently an input can only be attached to one channel).

describeInputResponse_httpStatus :: Lens' DescribeInputResponse Int

The response's http status code.