Network.AWS.MediaLive.DescribeInputDeviceThumbnail

Description

Get the latest thumbnail data for the input device.

Creating a Request

data DescribeInputDeviceThumbnail

Placeholder documentation for DescribeInputDeviceThumbnailRequest

See: newDescribeInputDeviceThumbnail smart constructor.

Constructors

DescribeInputDeviceThumbnail'

Fields

inputDeviceId :: Text

The unique ID of this input device. For example, hd-123456789abcdef.

accept :: AcceptHeader

The HTTP Accept header. Indicates the requested type for the thumbnail.

newDescribeInputDeviceThumbnail

Arguments

:: Text
$sel:inputDeviceId:DescribeInputDeviceThumbnail'

-> AcceptHeader
$sel:accept:DescribeInputDeviceThumbnail'

-> DescribeInputDeviceThumbnail

Create a value of DescribeInputDeviceThumbnail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputDeviceId:DescribeInputDeviceThumbnail', describeInputDeviceThumbnail_inputDeviceId - The unique ID of this input device. For example, hd-123456789abcdef.

$sel:accept:DescribeInputDeviceThumbnail', describeInputDeviceThumbnail_accept - The HTTP Accept header. Indicates the requested type for the thumbnail.

Request Lenses

describeInputDeviceThumbnail_inputDeviceId :: Lens' DescribeInputDeviceThumbnail Text

The unique ID of this input device. For example, hd-123456789abcdef.

describeInputDeviceThumbnail_accept :: Lens' DescribeInputDeviceThumbnail AcceptHeader

The HTTP Accept header. Indicates the requested type for the thumbnail.

Destructuring the Response

data DescribeInputDeviceThumbnailResponse

Placeholder documentation for DescribeInputDeviceThumbnailResponse

See: newDescribeInputDeviceThumbnailResponse smart constructor.

Constructors

DescribeInputDeviceThumbnailResponse'

Fields

eTag :: Maybe Text

The unique, cacheable version of this thumbnail.

contentLength :: Maybe Integer

The length of the content.

lastModified :: Maybe POSIX

The date and time the thumbnail was last updated at the device.

contentType :: Maybe ContentType

Specifies the media type of the thumbnail.

httpStatus :: Int

The response's http status code.

body :: ResponseBody

The binary data for the thumbnail that the Link device has most recently sent to MediaLive.

newDescribeInputDeviceThumbnailResponse

Arguments

:: Int
$sel:httpStatus:DescribeInputDeviceThumbnailResponse'

-> ResponseBody
$sel:body:DescribeInputDeviceThumbnailResponse'

-> DescribeInputDeviceThumbnailResponse

Create a value of DescribeInputDeviceThumbnailResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:DescribeInputDeviceThumbnailResponse', describeInputDeviceThumbnailResponse_eTag - The unique, cacheable version of this thumbnail.

$sel:contentLength:DescribeInputDeviceThumbnailResponse', describeInputDeviceThumbnailResponse_contentLength - The length of the content.

$sel:lastModified:DescribeInputDeviceThumbnailResponse', describeInputDeviceThumbnailResponse_lastModified - The date and time the thumbnail was last updated at the device.

$sel:contentType:DescribeInputDeviceThumbnailResponse', describeInputDeviceThumbnailResponse_contentType - Specifies the media type of the thumbnail.

$sel:httpStatus:DescribeInputDeviceThumbnailResponse', describeInputDeviceThumbnailResponse_httpStatus - The response's http status code.

$sel:body:DescribeInputDeviceThumbnailResponse', describeInputDeviceThumbnailResponse_body - The binary data for the thumbnail that the Link device has most recently sent to MediaLive.

Response Lenses

describeInputDeviceThumbnailResponse_eTag :: Lens' DescribeInputDeviceThumbnailResponse (Maybe Text)

The unique, cacheable version of this thumbnail.

describeInputDeviceThumbnailResponse_contentLength :: Lens' DescribeInputDeviceThumbnailResponse (Maybe Integer)

The length of the content.

describeInputDeviceThumbnailResponse_lastModified :: Lens' DescribeInputDeviceThumbnailResponse (Maybe UTCTime)

The date and time the thumbnail was last updated at the device.

describeInputDeviceThumbnailResponse_contentType :: Lens' DescribeInputDeviceThumbnailResponse (Maybe ContentType)

Specifies the media type of the thumbnail.

describeInputDeviceThumbnailResponse_httpStatus :: Lens' DescribeInputDeviceThumbnailResponse Int

The response's http status code.

describeInputDeviceThumbnailResponse_body :: Lens' DescribeInputDeviceThumbnailResponse ResponseBody

The binary data for the thumbnail that the Link device has most recently sent to MediaLive.