Network.AWS.MediaLive.DescribeInputSecurityGroup

Description

Produces a summary of an Input Security Group

Creating a Request

data DescribeInputSecurityGroup

Placeholder documentation for DescribeInputSecurityGroupRequest

See: newDescribeInputSecurityGroup smart constructor.

Constructors

DescribeInputSecurityGroup'

Fields

inputSecurityGroupId :: Text

The id of the Input Security Group to describe

newDescribeInputSecurityGroup

Arguments

:: Text
$sel:inputSecurityGroupId:DescribeInputSecurityGroup'

-> DescribeInputSecurityGroup

Create a value of DescribeInputSecurityGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputSecurityGroupId:DescribeInputSecurityGroup', describeInputSecurityGroup_inputSecurityGroupId - The id of the Input Security Group to describe

Request Lenses

describeInputSecurityGroup_inputSecurityGroupId :: Lens' DescribeInputSecurityGroup Text

The id of the Input Security Group to describe

Destructuring the Response

data DescribeInputSecurityGroupResponse

Placeholder documentation for DescribeInputSecurityGroupResponse

See: newDescribeInputSecurityGroupResponse smart constructor.

Constructors

DescribeInputSecurityGroupResponse'

Fields

arn :: Maybe Text

Unique ARN of Input Security Group

id :: Maybe Text

The Id of the Input Security Group

state :: Maybe InputSecurityGroupState

The current state of the Input Security Group.

tags :: Maybe (HashMap Text Text)

A collection of key-value pairs.

whitelistRules :: Maybe [InputWhitelistRule]

Whitelist rules and their sync status

inputs :: Maybe [Text]

The list of inputs currently using this Input Security Group.

httpStatus :: Int

The response's http status code.

newDescribeInputSecurityGroupResponse

Arguments

:: Int
$sel:httpStatus:DescribeInputSecurityGroupResponse'

-> DescribeInputSecurityGroupResponse

Create a value of DescribeInputSecurityGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:DescribeInputSecurityGroupResponse', describeInputSecurityGroupResponse_arn - Unique ARN of Input Security Group

$sel:id:DescribeInputSecurityGroupResponse', describeInputSecurityGroupResponse_id - The Id of the Input Security Group

$sel:state:DescribeInputSecurityGroupResponse', describeInputSecurityGroupResponse_state - The current state of the Input Security Group.

$sel:tags:DescribeInputSecurityGroupResponse', describeInputSecurityGroupResponse_tags - A collection of key-value pairs.

$sel:whitelistRules:DescribeInputSecurityGroupResponse', describeInputSecurityGroupResponse_whitelistRules - Whitelist rules and their sync status

$sel:inputs:DescribeInputSecurityGroupResponse', describeInputSecurityGroupResponse_inputs - The list of inputs currently using this Input Security Group.

$sel:httpStatus:DescribeInputSecurityGroupResponse', describeInputSecurityGroupResponse_httpStatus - The response's http status code.

Response Lenses

describeInputSecurityGroupResponse_arn :: Lens' DescribeInputSecurityGroupResponse (Maybe Text)

Unique ARN of Input Security Group

describeInputSecurityGroupResponse_id :: Lens' DescribeInputSecurityGroupResponse (Maybe Text)

The Id of the Input Security Group

describeInputSecurityGroupResponse_state :: Lens' DescribeInputSecurityGroupResponse (Maybe InputSecurityGroupState)

The current state of the Input Security Group.

describeInputSecurityGroupResponse_tags :: Lens' DescribeInputSecurityGroupResponse (Maybe (HashMap Text Text))

A collection of key-value pairs.

describeInputSecurityGroupResponse_whitelistRules :: Lens' DescribeInputSecurityGroupResponse (Maybe [InputWhitelistRule])

Whitelist rules and their sync status

describeInputSecurityGroupResponse_inputs :: Lens' DescribeInputSecurityGroupResponse (Maybe [Text])

The list of inputs currently using this Input Security Group.

describeInputSecurityGroupResponse_httpStatus :: Lens' DescribeInputSecurityGroupResponse Int

The response's http status code.