Network.AWS.MediaLive.DescribeMultiplex

Description

Gets details about a multiplex.

Creating a Request

data DescribeMultiplex

Placeholder documentation for DescribeMultiplexRequest

See: newDescribeMultiplex smart constructor.

Constructors

DescribeMultiplex'

Fields

multiplexId :: Text

The ID of the multiplex.

newDescribeMultiplex

Arguments

:: Text
$sel:multiplexId:DescribeMultiplex'

-> DescribeMultiplex

Create a value of DescribeMultiplex with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:multiplexId:DescribeMultiplex', describeMultiplex_multiplexId - The ID of the multiplex.

Request Lenses

describeMultiplex_multiplexId :: Lens' DescribeMultiplex Text

The ID of the multiplex.

Destructuring the Response

data DescribeMultiplexResponse

Placeholder documentation for DescribeMultiplexResponse

See: newDescribeMultiplexResponse smart constructor.

Constructors

DescribeMultiplexResponse'

Fields

state :: Maybe MultiplexState

The current state of the multiplex.

arn :: Maybe Text

The unique arn of the multiplex.

pipelinesRunningCount :: Maybe Int

The number of currently healthy pipelines.

availabilityZones :: Maybe [Text]

A list of availability zones for the multiplex.

programCount :: Maybe Int

The number of programs in the multiplex.

destinations :: Maybe [MultiplexOutputDestination]

A list of the multiplex output destinations.

name :: Maybe Text

The name of the multiplex.

id :: Maybe Text

The unique id of the multiplex.

multiplexSettings :: Maybe MultiplexSettings

Configuration for a multiplex event.

tags :: Maybe (HashMap Text Text)

A collection of key-value pairs.

httpStatus :: Int

The response's http status code.

newDescribeMultiplexResponse

Arguments

:: Int
$sel:httpStatus:DescribeMultiplexResponse'

-> DescribeMultiplexResponse

Create a value of DescribeMultiplexResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:DescribeMultiplexResponse', describeMultiplexResponse_state - The current state of the multiplex.

$sel:arn:DescribeMultiplexResponse', describeMultiplexResponse_arn - The unique arn of the multiplex.

$sel:pipelinesRunningCount:DescribeMultiplexResponse', describeMultiplexResponse_pipelinesRunningCount - The number of currently healthy pipelines.

$sel:availabilityZones:DescribeMultiplexResponse', describeMultiplexResponse_availabilityZones - A list of availability zones for the multiplex.

$sel:programCount:DescribeMultiplexResponse', describeMultiplexResponse_programCount - The number of programs in the multiplex.

$sel:destinations:DescribeMultiplexResponse', describeMultiplexResponse_destinations - A list of the multiplex output destinations.

$sel:name:DescribeMultiplexResponse', describeMultiplexResponse_name - The name of the multiplex.

$sel:id:DescribeMultiplexResponse', describeMultiplexResponse_id - The unique id of the multiplex.

$sel:multiplexSettings:DescribeMultiplexResponse', describeMultiplexResponse_multiplexSettings - Configuration for a multiplex event.

$sel:tags:DescribeMultiplexResponse', describeMultiplexResponse_tags - A collection of key-value pairs.

$sel:httpStatus:DescribeMultiplexResponse', describeMultiplexResponse_httpStatus - The response's http status code.

Response Lenses

describeMultiplexResponse_state :: Lens' DescribeMultiplexResponse (Maybe MultiplexState)

The current state of the multiplex.

describeMultiplexResponse_arn :: Lens' DescribeMultiplexResponse (Maybe Text)

The unique arn of the multiplex.

describeMultiplexResponse_pipelinesRunningCount :: Lens' DescribeMultiplexResponse (Maybe Int)

The number of currently healthy pipelines.

describeMultiplexResponse_availabilityZones :: Lens' DescribeMultiplexResponse (Maybe [Text])

A list of availability zones for the multiplex.

describeMultiplexResponse_programCount :: Lens' DescribeMultiplexResponse (Maybe Int)

The number of programs in the multiplex.

describeMultiplexResponse_destinations :: Lens' DescribeMultiplexResponse (Maybe [MultiplexOutputDestination])

A list of the multiplex output destinations.

describeMultiplexResponse_name :: Lens' DescribeMultiplexResponse (Maybe Text)

The name of the multiplex.

describeMultiplexResponse_id :: Lens' DescribeMultiplexResponse (Maybe Text)

The unique id of the multiplex.

describeMultiplexResponse_multiplexSettings :: Lens' DescribeMultiplexResponse (Maybe MultiplexSettings)

Configuration for a multiplex event.

describeMultiplexResponse_tags :: Lens' DescribeMultiplexResponse (Maybe (HashMap Text Text))

A collection of key-value pairs.

describeMultiplexResponse_httpStatus :: Lens' DescribeMultiplexResponse Int

The response's http status code.