Network.AWS.MediaLive.DescribeOffering

Description

Get details for an offering.

Creating a Request

data DescribeOffering

Placeholder documentation for DescribeOfferingRequest

See: newDescribeOffering smart constructor.

Constructors

DescribeOffering'

Fields

offeringId :: Text

Unique offering ID, e.g. '87654321'

newDescribeOffering

Arguments

:: Text
$sel:offeringId:DescribeOffering'

-> DescribeOffering

Create a value of DescribeOffering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:offeringId:DescribeOffering', describeOffering_offeringId - Unique offering ID, e.g. '87654321'

Request Lenses

describeOffering_offeringId :: Lens' DescribeOffering Text

Unique offering ID, e.g. '87654321'

Destructuring the Response

data DescribeOfferingResponse

Placeholder documentation for DescribeOfferingResponse

See: newDescribeOfferingResponse smart constructor.

Constructors

DescribeOfferingResponse'

Fields

duration :: Maybe Int

Lease duration, e.g. '12'

durationUnits :: Maybe OfferingDurationUnits

Units for duration, e.g. 'MONTHS'

arn :: Maybe Text

Unique offering ARN, e.g. 'arn:aws:medialive:us-west-2:123456789012:offering:87654321'

offeringId :: Maybe Text

Unique offering ID, e.g. '87654321'

currencyCode :: Maybe Text

Currency code for usagePrice and fixedPrice in ISO-4217 format, e.g. 'USD'

resourceSpecification :: Maybe ReservationResourceSpecification

Resource configuration details

offeringDescription :: Maybe Text

Offering description, e.g. 'HD AVC output at 10-20 Mbps, 30 fps, and standard VQ in US West (Oregon)'

fixedPrice :: Maybe Double

One-time charge for each reserved resource, e.g. '0.0' for a NO_UPFRONT offering

usagePrice :: Maybe Double

Recurring usage charge for each reserved resource, e.g. '157.0'

offeringType :: Maybe OfferingType

Offering type, e.g. 'NO_UPFRONT'

region :: Maybe Text

AWS region, e.g. 'us-west-2'

httpStatus :: Int

The response's http status code.

newDescribeOfferingResponse

Arguments

:: Int
$sel:httpStatus:DescribeOfferingResponse'

-> DescribeOfferingResponse

Create a value of DescribeOfferingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:duration:DescribeOfferingResponse', describeOfferingResponse_duration - Lease duration, e.g. '12'

$sel:durationUnits:DescribeOfferingResponse', describeOfferingResponse_durationUnits - Units for duration, e.g. 'MONTHS'

$sel:arn:DescribeOfferingResponse', describeOfferingResponse_arn - Unique offering ARN, e.g. 'arn:aws:medialive:us-west-2:123456789012:offering:87654321'

$sel:offeringId:DescribeOffering', describeOfferingResponse_offeringId - Unique offering ID, e.g. '87654321'

$sel:currencyCode:DescribeOfferingResponse', describeOfferingResponse_currencyCode - Currency code for usagePrice and fixedPrice in ISO-4217 format, e.g. 'USD'

$sel:resourceSpecification:DescribeOfferingResponse', describeOfferingResponse_resourceSpecification - Resource configuration details

$sel:offeringDescription:DescribeOfferingResponse', describeOfferingResponse_offeringDescription - Offering description, e.g. 'HD AVC output at 10-20 Mbps, 30 fps, and standard VQ in US West (Oregon)'

$sel:fixedPrice:DescribeOfferingResponse', describeOfferingResponse_fixedPrice - One-time charge for each reserved resource, e.g. '0.0' for a NO_UPFRONT offering

$sel:usagePrice:DescribeOfferingResponse', describeOfferingResponse_usagePrice - Recurring usage charge for each reserved resource, e.g. '157.0'

$sel:offeringType:DescribeOfferingResponse', describeOfferingResponse_offeringType - Offering type, e.g. 'NO_UPFRONT'

$sel:region:DescribeOfferingResponse', describeOfferingResponse_region - AWS region, e.g. 'us-west-2'

$sel:httpStatus:DescribeOfferingResponse', describeOfferingResponse_httpStatus - The response's http status code.

Response Lenses

describeOfferingResponse_duration :: Lens' DescribeOfferingResponse (Maybe Int)

Lease duration, e.g. '12'

describeOfferingResponse_durationUnits :: Lens' DescribeOfferingResponse (Maybe OfferingDurationUnits)

Units for duration, e.g. 'MONTHS'

describeOfferingResponse_arn :: Lens' DescribeOfferingResponse (Maybe Text)

Unique offering ARN, e.g. 'arn:aws:medialive:us-west-2:123456789012:offering:87654321'

describeOfferingResponse_offeringId :: Lens' DescribeOfferingResponse (Maybe Text)

Unique offering ID, e.g. '87654321'

describeOfferingResponse_currencyCode :: Lens' DescribeOfferingResponse (Maybe Text)

Currency code for usagePrice and fixedPrice in ISO-4217 format, e.g. 'USD'

describeOfferingResponse_resourceSpecification :: Lens' DescribeOfferingResponse (Maybe ReservationResourceSpecification)

Resource configuration details

describeOfferingResponse_offeringDescription :: Lens' DescribeOfferingResponse (Maybe Text)

Offering description, e.g. 'HD AVC output at 10-20 Mbps, 30 fps, and standard VQ in US West (Oregon)'

describeOfferingResponse_fixedPrice :: Lens' DescribeOfferingResponse (Maybe Double)

One-time charge for each reserved resource, e.g. '0.0' for a NO_UPFRONT offering

describeOfferingResponse_usagePrice :: Lens' DescribeOfferingResponse (Maybe Double)

Recurring usage charge for each reserved resource, e.g. '157.0'

describeOfferingResponse_offeringType :: Lens' DescribeOfferingResponse (Maybe OfferingType)

Offering type, e.g. 'NO_UPFRONT'

describeOfferingResponse_region :: Lens' DescribeOfferingResponse (Maybe Text)

AWS region, e.g. 'us-west-2'

describeOfferingResponse_httpStatus :: Lens' DescribeOfferingResponse Int

The response's http status code.