Network.AWS.MediaLive.DescribeSchedule

Description

Get a channel schedule

This operation returns paginated results.

Creating a Request

data DescribeSchedule

Placeholder documentation for DescribeScheduleRequest

See: newDescribeSchedule smart constructor.

Constructors

DescribeSchedule'

Fields

nextToken :: Maybe Text

maxResults :: Maybe Natural

channelId :: Text

Id of the channel whose schedule is being updated.

newDescribeSchedule

Arguments

:: Text
$sel:channelId:DescribeSchedule'

-> DescribeSchedule

Create a value of DescribeSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeSchedule', describeSchedule_nextToken - Undocumented member.

$sel:maxResults:DescribeSchedule', describeSchedule_maxResults - Undocumented member.

$sel:channelId:DescribeSchedule', describeSchedule_channelId - Id of the channel whose schedule is being updated.

Request Lenses

describeSchedule_nextToken :: Lens' DescribeSchedule (Maybe Text)

Undocumented member.

describeSchedule_maxResults :: Lens' DescribeSchedule (Maybe Natural)

Undocumented member.

describeSchedule_channelId :: Lens' DescribeSchedule Text

Id of the channel whose schedule is being updated.

Destructuring the Response

data DescribeScheduleResponse

Placeholder documentation for DescribeScheduleResponse

See: newDescribeScheduleResponse smart constructor.

Constructors

DescribeScheduleResponse'

Fields

nextToken :: Maybe Text

The next token; for use in pagination.

scheduleActions :: Maybe [ScheduleAction]

The list of actions in the schedule.

httpStatus :: Int

The response's http status code.

newDescribeScheduleResponse

Arguments

:: Int
$sel:httpStatus:DescribeScheduleResponse'

-> DescribeScheduleResponse

Create a value of DescribeScheduleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeSchedule', describeScheduleResponse_nextToken - The next token; for use in pagination.

$sel:scheduleActions:DescribeScheduleResponse', describeScheduleResponse_scheduleActions - The list of actions in the schedule.

$sel:httpStatus:DescribeScheduleResponse', describeScheduleResponse_httpStatus - The response's http status code.

Response Lenses

describeScheduleResponse_nextToken :: Lens' DescribeScheduleResponse (Maybe Text)

The next token; for use in pagination.

describeScheduleResponse_scheduleActions :: Lens' DescribeScheduleResponse (Maybe [ScheduleAction])

The list of actions in the schedule.

describeScheduleResponse_httpStatus :: Lens' DescribeScheduleResponse Int

The response's http status code.