Network.AWS.MediaLive.ListInputDeviceTransfers

Description

List input devices that are currently being transferred. List input devices that you are transferring from your AWS account or input devices that another AWS account is transferring to you.

This operation returns paginated results.

Creating a Request

data ListInputDeviceTransfers

Placeholder documentation for ListInputDeviceTransfersRequest

See: newListInputDeviceTransfers smart constructor.

Constructors

ListInputDeviceTransfers'

Fields

nextToken :: Maybe Text

maxResults :: Maybe Natural

transferType :: Text

newListInputDeviceTransfers

Arguments

:: Text
$sel:transferType:ListInputDeviceTransfers'

-> ListInputDeviceTransfers

Create a value of ListInputDeviceTransfers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListInputDeviceTransfers', listInputDeviceTransfers_nextToken - Undocumented member.

$sel:maxResults:ListInputDeviceTransfers', listInputDeviceTransfers_maxResults - Undocumented member.

$sel:transferType:ListInputDeviceTransfers', listInputDeviceTransfers_transferType - Undocumented member.

Request Lenses

listInputDeviceTransfers_nextToken :: Lens' ListInputDeviceTransfers (Maybe Text)

Undocumented member.

listInputDeviceTransfers_maxResults :: Lens' ListInputDeviceTransfers (Maybe Natural)

Undocumented member.

listInputDeviceTransfers_transferType :: Lens' ListInputDeviceTransfers Text

Undocumented member.

Destructuring the Response

data ListInputDeviceTransfersResponse

Placeholder documentation for ListInputDeviceTransfersResponse

See: newListInputDeviceTransfersResponse smart constructor.

Constructors

ListInputDeviceTransfersResponse'

Fields

nextToken :: Maybe Text

A token to get additional list results.

inputDeviceTransfers :: Maybe [TransferringInputDeviceSummary]

The list of devices that you are transferring or are being transferred to you.

httpStatus :: Int

The response's http status code.

newListInputDeviceTransfersResponse

Arguments

:: Int
$sel:httpStatus:ListInputDeviceTransfersResponse'

-> ListInputDeviceTransfersResponse

Create a value of ListInputDeviceTransfersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListInputDeviceTransfers', listInputDeviceTransfersResponse_nextToken - A token to get additional list results.

$sel:inputDeviceTransfers:ListInputDeviceTransfersResponse', listInputDeviceTransfersResponse_inputDeviceTransfers - The list of devices that you are transferring or are being transferred to you.

$sel:httpStatus:ListInputDeviceTransfersResponse', listInputDeviceTransfersResponse_httpStatus - The response's http status code.

Response Lenses

listInputDeviceTransfersResponse_nextToken :: Lens' ListInputDeviceTransfersResponse (Maybe Text)

A token to get additional list results.

listInputDeviceTransfersResponse_inputDeviceTransfers :: Lens' ListInputDeviceTransfersResponse (Maybe [TransferringInputDeviceSummary])

The list of devices that you are transferring or are being transferred to you.

listInputDeviceTransfersResponse_httpStatus :: Lens' ListInputDeviceTransfersResponse Int

The response's http status code.