Network.AWS.MediaLive.ListMultiplexPrograms

Description

List the programs that currently exist for a specific multiplex.

This operation returns paginated results.

Creating a Request

data ListMultiplexPrograms

Placeholder documentation for ListMultiplexProgramsRequest

See: newListMultiplexPrograms smart constructor.

Constructors

ListMultiplexPrograms'

Fields

nextToken :: Maybe Text

The token to retrieve the next page of results.

maxResults :: Maybe Natural

The maximum number of items to return.

multiplexId :: Text

The ID of the multiplex that the programs belong to.

newListMultiplexPrograms

Arguments

:: Text
$sel:multiplexId:ListMultiplexPrograms'

-> ListMultiplexPrograms

Create a value of ListMultiplexPrograms with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListMultiplexPrograms', listMultiplexPrograms_nextToken - The token to retrieve the next page of results.

$sel:maxResults:ListMultiplexPrograms', listMultiplexPrograms_maxResults - The maximum number of items to return.

$sel:multiplexId:ListMultiplexPrograms', listMultiplexPrograms_multiplexId - The ID of the multiplex that the programs belong to.

Request Lenses

listMultiplexPrograms_nextToken :: Lens' ListMultiplexPrograms (Maybe Text)

The token to retrieve the next page of results.

listMultiplexPrograms_maxResults :: Lens' ListMultiplexPrograms (Maybe Natural)

The maximum number of items to return.

listMultiplexPrograms_multiplexId :: Lens' ListMultiplexPrograms Text

The ID of the multiplex that the programs belong to.

Destructuring the Response

data ListMultiplexProgramsResponse

Placeholder documentation for ListMultiplexProgramsResponse

See: newListMultiplexProgramsResponse smart constructor.

Constructors

ListMultiplexProgramsResponse'

Fields

nextToken :: Maybe Text

Token for the next ListMultiplexProgram request.

multiplexPrograms :: Maybe [MultiplexProgramSummary]

List of multiplex programs.

httpStatus :: Int

The response's http status code.

newListMultiplexProgramsResponse

Arguments

:: Int
$sel:httpStatus:ListMultiplexProgramsResponse'

-> ListMultiplexProgramsResponse

Create a value of ListMultiplexProgramsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListMultiplexPrograms', listMultiplexProgramsResponse_nextToken - Token for the next ListMultiplexProgram request.

$sel:multiplexPrograms:ListMultiplexProgramsResponse', listMultiplexProgramsResponse_multiplexPrograms - List of multiplex programs.

$sel:httpStatus:ListMultiplexProgramsResponse', listMultiplexProgramsResponse_httpStatus - The response's http status code.

Response Lenses

listMultiplexProgramsResponse_nextToken :: Lens' ListMultiplexProgramsResponse (Maybe Text)

Token for the next ListMultiplexProgram request.

listMultiplexProgramsResponse_multiplexPrograms :: Lens' ListMultiplexProgramsResponse (Maybe [MultiplexProgramSummary])

List of multiplex programs.

listMultiplexProgramsResponse_httpStatus :: Lens' ListMultiplexProgramsResponse Int

The response's http status code.