Network.AWS.MediaLive.ListMultiplexes

Description

Retrieve a list of the existing multiplexes.

This operation returns paginated results.

Creating a Request

data ListMultiplexes

Placeholder documentation for ListMultiplexesRequest

See: newListMultiplexes smart constructor.

Constructors

ListMultiplexes'

Fields

nextToken :: Maybe Text

The token to retrieve the next page of results.

maxResults :: Maybe Natural

The maximum number of items to return.

newListMultiplexes :: ListMultiplexes

Create a value of ListMultiplexes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListMultiplexes', listMultiplexes_nextToken - The token to retrieve the next page of results.

$sel:maxResults:ListMultiplexes', listMultiplexes_maxResults - The maximum number of items to return.

Request Lenses

listMultiplexes_nextToken :: Lens' ListMultiplexes (Maybe Text)

The token to retrieve the next page of results.

listMultiplexes_maxResults :: Lens' ListMultiplexes (Maybe Natural)

The maximum number of items to return.

Destructuring the Response

data ListMultiplexesResponse

Placeholder documentation for ListMultiplexesResponse

See: newListMultiplexesResponse smart constructor.

Constructors

ListMultiplexesResponse'

Fields

nextToken :: Maybe Text

Token for the next ListMultiplexes request.

multiplexes :: Maybe [MultiplexSummary]

List of multiplexes.

httpStatus :: Int

The response's http status code.

newListMultiplexesResponse

Arguments

:: Int
$sel:httpStatus:ListMultiplexesResponse'

-> ListMultiplexesResponse

Create a value of ListMultiplexesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListMultiplexes', listMultiplexesResponse_nextToken - Token for the next ListMultiplexes request.

$sel:multiplexes:ListMultiplexesResponse', listMultiplexesResponse_multiplexes - List of multiplexes.

$sel:httpStatus:ListMultiplexesResponse', listMultiplexesResponse_httpStatus - The response's http status code.

Response Lenses

listMultiplexesResponse_nextToken :: Lens' ListMultiplexesResponse (Maybe Text)

Token for the next ListMultiplexes request.

listMultiplexesResponse_multiplexes :: Lens' ListMultiplexesResponse (Maybe [MultiplexSummary])

List of multiplexes.

listMultiplexesResponse_httpStatus :: Lens' ListMultiplexesResponse Int

The response's http status code.