Network.AWS.MediaLive.ListReservations

Description

List purchased reservations.

This operation returns paginated results.

Creating a Request

data ListReservations

Placeholder documentation for ListReservationsRequest

See: newListReservations smart constructor.

Constructors

ListReservations'

Fields

maximumFramerate :: Maybe Text

Filter by framerate, 'MAX_30_FPS' or 'MAX_60_FPS'

nextToken :: Maybe Text

videoQuality :: Maybe Text

Filter by video quality, 'STANDARD', 'ENHANCED', or 'PREMIUM'

maxResults :: Maybe Natural

codec :: Maybe Text

Filter by codec, 'AVC', 'HEVC', 'MPEG2', 'AUDIO', or 'LINK'

maximumBitrate :: Maybe Text

Filter by bitrate, 'MAX_10_MBPS', 'MAX_20_MBPS', or 'MAX_50_MBPS'

channelClass :: Maybe Text

Filter by channel class, 'STANDARD' or 'SINGLE_PIPELINE'

specialFeature :: Maybe Text

Filter by special feature, 'ADVANCED_AUDIO' or 'AUDIO_NORMALIZATION'

resourceType :: Maybe Text

Filter by resource type, 'INPUT', 'OUTPUT', 'MULTIPLEX', or 'CHANNEL'

resolution :: Maybe Text

Filter by resolution, 'SD', 'HD', 'FHD', or 'UHD'

newListReservations :: ListReservations

Create a value of ListReservations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maximumFramerate:ListReservations', listReservations_maximumFramerate - Filter by framerate, 'MAX_30_FPS' or 'MAX_60_FPS'

$sel:nextToken:ListReservations', listReservations_nextToken - Undocumented member.

$sel:videoQuality:ListReservations', listReservations_videoQuality - Filter by video quality, 'STANDARD', 'ENHANCED', or 'PREMIUM'

$sel:maxResults:ListReservations', listReservations_maxResults - Undocumented member.

$sel:codec:ListReservations', listReservations_codec - Filter by codec, 'AVC', 'HEVC', 'MPEG2', 'AUDIO', or 'LINK'

$sel:maximumBitrate:ListReservations', listReservations_maximumBitrate - Filter by bitrate, 'MAX_10_MBPS', 'MAX_20_MBPS', or 'MAX_50_MBPS'

$sel:channelClass:ListReservations', listReservations_channelClass - Filter by channel class, 'STANDARD' or 'SINGLE_PIPELINE'

$sel:specialFeature:ListReservations', listReservations_specialFeature - Filter by special feature, 'ADVANCED_AUDIO' or 'AUDIO_NORMALIZATION'

$sel:resourceType:ListReservations', listReservations_resourceType - Filter by resource type, 'INPUT', 'OUTPUT', 'MULTIPLEX', or 'CHANNEL'

$sel:resolution:ListReservations', listReservations_resolution - Filter by resolution, 'SD', 'HD', 'FHD', or 'UHD'

Request Lenses

listReservations_maximumFramerate :: Lens' ListReservations (Maybe Text)

Filter by framerate, 'MAX_30_FPS' or 'MAX_60_FPS'

listReservations_nextToken :: Lens' ListReservations (Maybe Text)

Undocumented member.

listReservations_videoQuality :: Lens' ListReservations (Maybe Text)

Filter by video quality, 'STANDARD', 'ENHANCED', or 'PREMIUM'

listReservations_maxResults :: Lens' ListReservations (Maybe Natural)

Undocumented member.

listReservations_codec :: Lens' ListReservations (Maybe Text)

Filter by codec, 'AVC', 'HEVC', 'MPEG2', 'AUDIO', or 'LINK'

listReservations_maximumBitrate :: Lens' ListReservations (Maybe Text)

Filter by bitrate, 'MAX_10_MBPS', 'MAX_20_MBPS', or 'MAX_50_MBPS'

listReservations_channelClass :: Lens' ListReservations (Maybe Text)

Filter by channel class, 'STANDARD' or 'SINGLE_PIPELINE'

listReservations_specialFeature :: Lens' ListReservations (Maybe Text)

Filter by special feature, 'ADVANCED_AUDIO' or 'AUDIO_NORMALIZATION'

listReservations_resourceType :: Lens' ListReservations (Maybe Text)

Filter by resource type, 'INPUT', 'OUTPUT', 'MULTIPLEX', or 'CHANNEL'

listReservations_resolution :: Lens' ListReservations (Maybe Text)

Filter by resolution, 'SD', 'HD', 'FHD', or 'UHD'

Destructuring the Response

data ListReservationsResponse

Placeholder documentation for ListReservationsResponse

See: newListReservationsResponse smart constructor.

Constructors

ListReservationsResponse'

Fields

nextToken :: Maybe Text

Token to retrieve the next page of results

reservations :: Maybe [Reservation]

List of reservations

httpStatus :: Int

The response's http status code.

newListReservationsResponse

Arguments

:: Int
$sel:httpStatus:ListReservationsResponse'

-> ListReservationsResponse

Create a value of ListReservationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListReservations', listReservationsResponse_nextToken - Token to retrieve the next page of results

$sel:reservations:ListReservationsResponse', listReservationsResponse_reservations - List of reservations

$sel:httpStatus:ListReservationsResponse', listReservationsResponse_httpStatus - The response's http status code.

Response Lenses

listReservationsResponse_nextToken :: Lens' ListReservationsResponse (Maybe Text)

Token to retrieve the next page of results

listReservationsResponse_reservations :: Lens' ListReservationsResponse (Maybe [Reservation])

List of reservations

listReservationsResponse_httpStatus :: Lens' ListReservationsResponse Int

The response's http status code.