Network.AWS.MediaLive.StartChannel

Description

Starts an existing channel

Creating a Request

data StartChannel

Placeholder documentation for StartChannelRequest

See: newStartChannel smart constructor.

Constructors

StartChannel'

Fields

channelId :: Text

A request to start a channel

newStartChannel

Arguments

:: Text
$sel:channelId:StartChannel'

-> StartChannel

Create a value of StartChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelId:StartChannel', startChannel_channelId - A request to start a channel

Request Lenses

startChannel_channelId :: Lens' StartChannel Text

A request to start a channel

Destructuring the Response

data StartChannelResponse

Placeholder documentation for StartChannelResponse

See: newStartChannelResponse smart constructor.

Constructors

StartChannelResponse'

Fields

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the role assumed when running the Channel.

encoderSettings :: Maybe EncoderSettings

inputSpecification :: Maybe InputSpecification

Specification of network and file inputs for this channel

arn :: Maybe Text

The unique arn of the channel.

id :: Maybe Text

The unique id of the channel.

pipelinesRunningCount :: Maybe Int

The number of currently healthy pipelines.

channelClass :: Maybe ChannelClass

The class for this channel. STANDARD for a channel with two pipelines or SINGLE_PIPELINE for a channel with one pipeline.

logLevel :: Maybe LogLevel

The log level being written to CloudWatch Logs.

name :: Maybe Text

The name of the channel. (user-mutable)

destinations :: Maybe [OutputDestination]

A list of destinations of the channel. For UDP outputs, there is one destination per output. For other types (HLS, for example), there is one destination per packager.

state :: Maybe ChannelState

inputAttachments :: Maybe [InputAttachment]

List of input attachments for channel.

tags :: Maybe (HashMap Text Text)

A collection of key-value pairs.

pipelineDetails :: Maybe [PipelineDetail]

Runtime details for the pipelines of a running channel.

vpc :: Maybe VpcOutputSettingsDescription

Settings for VPC output

cdiInputSpecification :: Maybe CdiInputSpecification

Specification of CDI inputs for this channel

egressEndpoints :: Maybe [ChannelEgressEndpoint]

The endpoints where outgoing connections initiate from

httpStatus :: Int

The response's http status code.

newStartChannelResponse

Arguments

:: Int
$sel:httpStatus:StartChannelResponse'

-> StartChannelResponse

Create a value of StartChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleArn:StartChannelResponse', startChannelResponse_roleArn - The Amazon Resource Name (ARN) of the role assumed when running the Channel.

$sel:encoderSettings:StartChannelResponse', startChannelResponse_encoderSettings - Undocumented member.

$sel:inputSpecification:StartChannelResponse', startChannelResponse_inputSpecification - Specification of network and file inputs for this channel

$sel:arn:StartChannelResponse', startChannelResponse_arn - The unique arn of the channel.

$sel:id:StartChannelResponse', startChannelResponse_id - The unique id of the channel.

$sel:pipelinesRunningCount:StartChannelResponse', startChannelResponse_pipelinesRunningCount - The number of currently healthy pipelines.

$sel:channelClass:StartChannelResponse', startChannelResponse_channelClass - The class for this channel. STANDARD for a channel with two pipelines or SINGLE_PIPELINE for a channel with one pipeline.

$sel:logLevel:StartChannelResponse', startChannelResponse_logLevel - The log level being written to CloudWatch Logs.

$sel:name:StartChannelResponse', startChannelResponse_name - The name of the channel. (user-mutable)

$sel:destinations:StartChannelResponse', startChannelResponse_destinations - A list of destinations of the channel. For UDP outputs, there is one destination per output. For other types (HLS, for example), there is one destination per packager.

$sel:state:StartChannelResponse', startChannelResponse_state - Undocumented member.

$sel:inputAttachments:StartChannelResponse', startChannelResponse_inputAttachments - List of input attachments for channel.

$sel:tags:StartChannelResponse', startChannelResponse_tags - A collection of key-value pairs.

$sel:pipelineDetails:StartChannelResponse', startChannelResponse_pipelineDetails - Runtime details for the pipelines of a running channel.

$sel:vpc:StartChannelResponse', startChannelResponse_vpc - Settings for VPC output

$sel:cdiInputSpecification:StartChannelResponse', startChannelResponse_cdiInputSpecification - Specification of CDI inputs for this channel

$sel:egressEndpoints:StartChannelResponse', startChannelResponse_egressEndpoints - The endpoints where outgoing connections initiate from

$sel:httpStatus:StartChannelResponse', startChannelResponse_httpStatus - The response's http status code.

Response Lenses

startChannelResponse_roleArn :: Lens' StartChannelResponse (Maybe Text)

The Amazon Resource Name (ARN) of the role assumed when running the Channel.

startChannelResponse_encoderSettings :: Lens' StartChannelResponse (Maybe EncoderSettings)

Undocumented member.

startChannelResponse_inputSpecification :: Lens' StartChannelResponse (Maybe InputSpecification)

Specification of network and file inputs for this channel

startChannelResponse_arn :: Lens' StartChannelResponse (Maybe Text)

The unique arn of the channel.

startChannelResponse_id :: Lens' StartChannelResponse (Maybe Text)

The unique id of the channel.

startChannelResponse_pipelinesRunningCount :: Lens' StartChannelResponse (Maybe Int)

The number of currently healthy pipelines.

startChannelResponse_channelClass :: Lens' StartChannelResponse (Maybe ChannelClass)

The class for this channel. STANDARD for a channel with two pipelines or SINGLE_PIPELINE for a channel with one pipeline.

startChannelResponse_logLevel :: Lens' StartChannelResponse (Maybe LogLevel)

The log level being written to CloudWatch Logs.

startChannelResponse_name :: Lens' StartChannelResponse (Maybe Text)

The name of the channel. (user-mutable)

startChannelResponse_destinations :: Lens' StartChannelResponse (Maybe [OutputDestination])

A list of destinations of the channel. For UDP outputs, there is one destination per output. For other types (HLS, for example), there is one destination per packager.

startChannelResponse_state :: Lens' StartChannelResponse (Maybe ChannelState)

Undocumented member.

startChannelResponse_inputAttachments :: Lens' StartChannelResponse (Maybe [InputAttachment])

List of input attachments for channel.

startChannelResponse_tags :: Lens' StartChannelResponse (Maybe (HashMap Text Text))

A collection of key-value pairs.

startChannelResponse_pipelineDetails :: Lens' StartChannelResponse (Maybe [PipelineDetail])

Runtime details for the pipelines of a running channel.

startChannelResponse_vpc :: Lens' StartChannelResponse (Maybe VpcOutputSettingsDescription)

Settings for VPC output

startChannelResponse_cdiInputSpecification :: Lens' StartChannelResponse (Maybe CdiInputSpecification)

Specification of CDI inputs for this channel

startChannelResponse_egressEndpoints :: Lens' StartChannelResponse (Maybe [ChannelEgressEndpoint])

The endpoints where outgoing connections initiate from

startChannelResponse_httpStatus :: Lens' StartChannelResponse Int

The response's http status code.