Network.AWS.MediaLive.StartMultiplex

Description

Start (run) the multiplex. Starting the multiplex does not start the channels. You must explicitly start each channel.

Creating a Request

data StartMultiplex

Placeholder documentation for StartMultiplexRequest

See: newStartMultiplex smart constructor.

Constructors

StartMultiplex'

Fields

multiplexId :: Text

The ID of the multiplex.

newStartMultiplex

Arguments

:: Text
$sel:multiplexId:StartMultiplex'

-> StartMultiplex

Create a value of StartMultiplex with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:multiplexId:StartMultiplex', startMultiplex_multiplexId - The ID of the multiplex.

Request Lenses

startMultiplex_multiplexId :: Lens' StartMultiplex Text

The ID of the multiplex.

Destructuring the Response

data StartMultiplexResponse

Placeholder documentation for StartMultiplexResponse

See: newStartMultiplexResponse smart constructor.

Constructors

StartMultiplexResponse'

Fields

availabilityZones :: Maybe [Text]

A list of availability zones for the multiplex.

arn :: Maybe Text

The unique arn of the multiplex.

id :: Maybe Text

The unique id of the multiplex.

pipelinesRunningCount :: Maybe Int

The number of currently healthy pipelines.

name :: Maybe Text

The name of the multiplex.

destinations :: Maybe [MultiplexOutputDestination]

A list of the multiplex output destinations.

programCount :: Maybe Int

The number of programs in the multiplex.

state :: Maybe MultiplexState

The current state of the multiplex.

tags :: Maybe (HashMap Text Text)

A collection of key-value pairs.

multiplexSettings :: Maybe MultiplexSettings

Configuration for a multiplex event.

httpStatus :: Int

The response's http status code.

newStartMultiplexResponse

Arguments

:: Int
$sel:httpStatus:StartMultiplexResponse'

-> StartMultiplexResponse

Create a value of StartMultiplexResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availabilityZones:StartMultiplexResponse', startMultiplexResponse_availabilityZones - A list of availability zones for the multiplex.

$sel:arn:StartMultiplexResponse', startMultiplexResponse_arn - The unique arn of the multiplex.

$sel:id:StartMultiplexResponse', startMultiplexResponse_id - The unique id of the multiplex.

$sel:pipelinesRunningCount:StartMultiplexResponse', startMultiplexResponse_pipelinesRunningCount - The number of currently healthy pipelines.

$sel:name:StartMultiplexResponse', startMultiplexResponse_name - The name of the multiplex.

$sel:destinations:StartMultiplexResponse', startMultiplexResponse_destinations - A list of the multiplex output destinations.

$sel:programCount:StartMultiplexResponse', startMultiplexResponse_programCount - The number of programs in the multiplex.

$sel:state:StartMultiplexResponse', startMultiplexResponse_state - The current state of the multiplex.

$sel:tags:StartMultiplexResponse', startMultiplexResponse_tags - A collection of key-value pairs.

$sel:multiplexSettings:StartMultiplexResponse', startMultiplexResponse_multiplexSettings - Configuration for a multiplex event.

$sel:httpStatus:StartMultiplexResponse', startMultiplexResponse_httpStatus - The response's http status code.

Response Lenses

startMultiplexResponse_availabilityZones :: Lens' StartMultiplexResponse (Maybe [Text])

A list of availability zones for the multiplex.

startMultiplexResponse_arn :: Lens' StartMultiplexResponse (Maybe Text)

The unique arn of the multiplex.

startMultiplexResponse_id :: Lens' StartMultiplexResponse (Maybe Text)

The unique id of the multiplex.

startMultiplexResponse_pipelinesRunningCount :: Lens' StartMultiplexResponse (Maybe Int)

The number of currently healthy pipelines.

startMultiplexResponse_name :: Lens' StartMultiplexResponse (Maybe Text)

The name of the multiplex.

startMultiplexResponse_destinations :: Lens' StartMultiplexResponse (Maybe [MultiplexOutputDestination])

A list of the multiplex output destinations.

startMultiplexResponse_programCount :: Lens' StartMultiplexResponse (Maybe Int)

The number of programs in the multiplex.

startMultiplexResponse_state :: Lens' StartMultiplexResponse (Maybe MultiplexState)

The current state of the multiplex.

startMultiplexResponse_tags :: Lens' StartMultiplexResponse (Maybe (HashMap Text Text))

A collection of key-value pairs.

startMultiplexResponse_multiplexSettings :: Lens' StartMultiplexResponse (Maybe MultiplexSettings)

Configuration for a multiplex event.

startMultiplexResponse_httpStatus :: Lens' StartMultiplexResponse Int

The response's http status code.