Network.AWS.MediaLive.StopMultiplex

Description

Stops a running multiplex. If the multiplex isn't running, this action has no effect.

Creating a Request

data StopMultiplex

Placeholder documentation for StopMultiplexRequest

See: newStopMultiplex smart constructor.

Constructors

StopMultiplex'

Fields

multiplexId :: Text

The ID of the multiplex.

newStopMultiplex

Arguments

:: Text
$sel:multiplexId:StopMultiplex'

-> StopMultiplex

Create a value of StopMultiplex with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:multiplexId:StopMultiplex', stopMultiplex_multiplexId - The ID of the multiplex.

Request Lenses

stopMultiplex_multiplexId :: Lens' StopMultiplex Text

The ID of the multiplex.

Destructuring the Response

data StopMultiplexResponse

Placeholder documentation for StopMultiplexResponse

See: newStopMultiplexResponse smart constructor.

Constructors

StopMultiplexResponse'

Fields

availabilityZones :: Maybe [Text]

A list of availability zones for the multiplex.

arn :: Maybe Text

The unique arn of the multiplex.

id :: Maybe Text

The unique id of the multiplex.

pipelinesRunningCount :: Maybe Int

The number of currently healthy pipelines.

name :: Maybe Text

The name of the multiplex.

destinations :: Maybe [MultiplexOutputDestination]

A list of the multiplex output destinations.

programCount :: Maybe Int

The number of programs in the multiplex.

state :: Maybe MultiplexState

The current state of the multiplex.

tags :: Maybe (HashMap Text Text)

A collection of key-value pairs.

multiplexSettings :: Maybe MultiplexSettings

Configuration for a multiplex event.

httpStatus :: Int

The response's http status code.

newStopMultiplexResponse

Arguments

:: Int
$sel:httpStatus:StopMultiplexResponse'

-> StopMultiplexResponse

Create a value of StopMultiplexResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availabilityZones:StopMultiplexResponse', stopMultiplexResponse_availabilityZones - A list of availability zones for the multiplex.

$sel:arn:StopMultiplexResponse', stopMultiplexResponse_arn - The unique arn of the multiplex.

$sel:id:StopMultiplexResponse', stopMultiplexResponse_id - The unique id of the multiplex.

$sel:pipelinesRunningCount:StopMultiplexResponse', stopMultiplexResponse_pipelinesRunningCount - The number of currently healthy pipelines.

$sel:name:StopMultiplexResponse', stopMultiplexResponse_name - The name of the multiplex.

$sel:destinations:StopMultiplexResponse', stopMultiplexResponse_destinations - A list of the multiplex output destinations.

$sel:programCount:StopMultiplexResponse', stopMultiplexResponse_programCount - The number of programs in the multiplex.

$sel:state:StopMultiplexResponse', stopMultiplexResponse_state - The current state of the multiplex.

$sel:tags:StopMultiplexResponse', stopMultiplexResponse_tags - A collection of key-value pairs.

$sel:multiplexSettings:StopMultiplexResponse', stopMultiplexResponse_multiplexSettings - Configuration for a multiplex event.

$sel:httpStatus:StopMultiplexResponse', stopMultiplexResponse_httpStatus - The response's http status code.

Response Lenses

stopMultiplexResponse_availabilityZones :: Lens' StopMultiplexResponse (Maybe [Text])

A list of availability zones for the multiplex.

stopMultiplexResponse_arn :: Lens' StopMultiplexResponse (Maybe Text)

The unique arn of the multiplex.

stopMultiplexResponse_id :: Lens' StopMultiplexResponse (Maybe Text)

The unique id of the multiplex.

stopMultiplexResponse_pipelinesRunningCount :: Lens' StopMultiplexResponse (Maybe Int)

The number of currently healthy pipelines.

stopMultiplexResponse_name :: Lens' StopMultiplexResponse (Maybe Text)

The name of the multiplex.

stopMultiplexResponse_destinations :: Lens' StopMultiplexResponse (Maybe [MultiplexOutputDestination])

A list of the multiplex output destinations.

stopMultiplexResponse_programCount :: Lens' StopMultiplexResponse (Maybe Int)

The number of programs in the multiplex.

stopMultiplexResponse_state :: Lens' StopMultiplexResponse (Maybe MultiplexState)

The current state of the multiplex.

stopMultiplexResponse_tags :: Lens' StopMultiplexResponse (Maybe (HashMap Text Text))

A collection of key-value pairs.

stopMultiplexResponse_multiplexSettings :: Lens' StopMultiplexResponse (Maybe MultiplexSettings)

Configuration for a multiplex event.

stopMultiplexResponse_httpStatus :: Lens' StopMultiplexResponse Int

The response's http status code.