Network.AWS.MediaLive.TransferInputDevice

Description

Start an input device transfer to another AWS account. After you make the request, the other account must accept or reject the transfer.

Creating a Request

data TransferInputDevice'

A request to transfer an input device.

See: newTransferInputDevice' smart constructor.

Constructors

TransferInputDevice''

Fields

targetRegion :: Maybe Text

The target AWS region to transfer the device.

transferMessage :: Maybe Text

An optional message for the recipient. Maximum 280 characters.

targetCustomerId :: Maybe Text

The AWS account ID (12 digits) for the recipient of the device transfer.

inputDeviceId :: Text

The unique ID of this input device. For example, hd-123456789abcdef.

newTransferInputDevice'

Arguments

:: Text
$sel:inputDeviceId:TransferInputDevice''

-> TransferInputDevice'

Create a value of TransferInputDevice' with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetRegion:TransferInputDevice'', transferInputDevice'_targetRegion - The target AWS region to transfer the device.

$sel:transferMessage:TransferInputDevice'', transferInputDevice'_transferMessage - An optional message for the recipient. Maximum 280 characters.

$sel:targetCustomerId:TransferInputDevice'', transferInputDevice'_targetCustomerId - The AWS account ID (12 digits) for the recipient of the device transfer.

$sel:inputDeviceId:TransferInputDevice'', transferInputDevice'_inputDeviceId - The unique ID of this input device. For example, hd-123456789abcdef.

Request Lenses

transferInputDevice'_targetRegion :: Lens' TransferInputDevice' (Maybe Text)

The target AWS region to transfer the device.

transferInputDevice'_transferMessage :: Lens' TransferInputDevice' (Maybe Text)

An optional message for the recipient. Maximum 280 characters.

transferInputDevice'_targetCustomerId :: Lens' TransferInputDevice' (Maybe Text)

The AWS account ID (12 digits) for the recipient of the device transfer.

transferInputDevice'_inputDeviceId :: Lens' TransferInputDevice' Text

The unique ID of this input device. For example, hd-123456789abcdef.

Destructuring the Response

data TransferInputDeviceResponse

Placeholder documentation for TransferInputDeviceResponse

See: newTransferInputDeviceResponse smart constructor.

Constructors

TransferInputDeviceResponse'

Fields

httpStatus :: Int

The response's http status code.

newTransferInputDeviceResponse

Arguments

:: Int
$sel:httpStatus:TransferInputDeviceResponse'

-> TransferInputDeviceResponse

Create a value of TransferInputDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TransferInputDeviceResponse', transferInputDeviceResponse_httpStatus - The response's http status code.

Response Lenses

transferInputDeviceResponse_httpStatus :: Lens' TransferInputDeviceResponse Int

The response's http status code.