Network.AWS.MediaLive.Types.AacSettings

Description

Documentation

data AacSettings

Aac Settings

See: newAacSettings smart constructor.

Constructors

AacSettings'

Fields

rawFormat :: Maybe AacRawFormat

Sets LATM / LOAS AAC output for raw containers.

codingMode :: Maybe AacCodingMode

Mono, Stereo, or 5.1 channel layout. Valid values depend on rate control mode and profile. The adReceiverMix setting receives a stereo description plus control track and emits a mono AAC encode of the description track, with control data emitted in the PES header as per ETSI TS 101 154 Annex E.

profile :: Maybe AacProfile

AAC Profile.

rateControlMode :: Maybe AacRateControlMode

Rate Control Mode.

sampleRate :: Maybe Double

Sample rate in Hz. Valid values depend on rate control mode and profile.

spec :: Maybe AacSpec

Use MPEG-2 AAC audio instead of MPEG-4 AAC audio for raw or MPEG-2 Transport Stream containers.

bitrate :: Maybe Double

Average bitrate in bits/second. Valid values depend on rate control mode and profile.

vbrQuality :: Maybe AacVbrQuality

VBR Quality Level - Only used if rateControlMode is VBR.

inputType :: Maybe AacInputType

Set to "broadcasterMixedAd" when input contains pre-mixed main audio + AD (narration) as a stereo pair. The Audio Type field (audioType) will be set to 3, which signals to downstream systems that this stream contains "broadcaster mixed AD". Note that the input received by the encoder must contain pre-mixed audio; the encoder does not perform the mixing. The values in audioTypeControl and audioType (in AudioDescription) are ignored when set to broadcasterMixedAd. Leave set to "normal" when input does not contain pre-mixed audio + AD.

newAacSettings :: AacSettings

Create a value of AacSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rawFormat:AacSettings', aacSettings_rawFormat - Sets LATM / LOAS AAC output for raw containers.

$sel:codingMode:AacSettings', aacSettings_codingMode - Mono, Stereo, or 5.1 channel layout. Valid values depend on rate control mode and profile. The adReceiverMix setting receives a stereo description plus control track and emits a mono AAC encode of the description track, with control data emitted in the PES header as per ETSI TS 101 154 Annex E.

$sel:profile:AacSettings', aacSettings_profile - AAC Profile.

$sel:rateControlMode:AacSettings', aacSettings_rateControlMode - Rate Control Mode.

$sel:sampleRate:AacSettings', aacSettings_sampleRate - Sample rate in Hz. Valid values depend on rate control mode and profile.

$sel:spec:AacSettings', aacSettings_spec - Use MPEG-2 AAC audio instead of MPEG-4 AAC audio for raw or MPEG-2 Transport Stream containers.

$sel:bitrate:AacSettings', aacSettings_bitrate - Average bitrate in bits/second. Valid values depend on rate control mode and profile.

$sel:vbrQuality:AacSettings', aacSettings_vbrQuality - VBR Quality Level - Only used if rateControlMode is VBR.

$sel:inputType:AacSettings', aacSettings_inputType - Set to "broadcasterMixedAd" when input contains pre-mixed main audio + AD (narration) as a stereo pair. The Audio Type field (audioType) will be set to 3, which signals to downstream systems that this stream contains "broadcaster mixed AD". Note that the input received by the encoder must contain pre-mixed audio; the encoder does not perform the mixing. The values in audioTypeControl and audioType (in AudioDescription) are ignored when set to broadcasterMixedAd. Leave set to "normal" when input does not contain pre-mixed audio + AD.

aacSettings_rawFormat :: Lens' AacSettings (Maybe AacRawFormat)

Sets LATM / LOAS AAC output for raw containers.

aacSettings_codingMode :: Lens' AacSettings (Maybe AacCodingMode)

Mono, Stereo, or 5.1 channel layout. Valid values depend on rate control mode and profile. The adReceiverMix setting receives a stereo description plus control track and emits a mono AAC encode of the description track, with control data emitted in the PES header as per ETSI TS 101 154 Annex E.

aacSettings_profile :: Lens' AacSettings (Maybe AacProfile)

AAC Profile.

aacSettings_rateControlMode :: Lens' AacSettings (Maybe AacRateControlMode)

Rate Control Mode.

aacSettings_sampleRate :: Lens' AacSettings (Maybe Double)

Sample rate in Hz. Valid values depend on rate control mode and profile.

aacSettings_spec :: Lens' AacSettings (Maybe AacSpec)

Use MPEG-2 AAC audio instead of MPEG-4 AAC audio for raw or MPEG-2 Transport Stream containers.

aacSettings_bitrate :: Lens' AacSettings (Maybe Double)

Average bitrate in bits/second. Valid values depend on rate control mode and profile.

aacSettings_vbrQuality :: Lens' AacSettings (Maybe AacVbrQuality)

VBR Quality Level - Only used if rateControlMode is VBR.

aacSettings_inputType :: Lens' AacSettings (Maybe AacInputType)

Set to "broadcasterMixedAd" when input contains pre-mixed main audio + AD (narration) as a stereo pair. The Audio Type field (audioType) will be set to 3, which signals to downstream systems that this stream contains "broadcaster mixed AD". Note that the input received by the encoder must contain pre-mixed audio; the encoder does not perform the mixing. The values in audioTypeControl and audioType (in AudioDescription) are ignored when set to broadcasterMixedAd. Leave set to "normal" when input does not contain pre-mixed audio + AD.