Network.AWS.MediaLive.Types.AudioDescription

Description

Documentation

data AudioDescription

Audio Description

See: newAudioDescription smart constructor.

Constructors

AudioDescription'

Fields

languageCode :: Maybe Text

RFC 5646 language code representing the language of the audio output track. Only used if languageControlMode is useConfigured, or there is no ISO 639 language code specified in the input.

audioType :: Maybe AudioType

Applies only if audioTypeControl is useConfigured. The values for audioType are defined in ISO-IEC 13818-1.

audioNormalizationSettings :: Maybe AudioNormalizationSettings

Advanced audio normalization settings.

languageCodeControl :: Maybe AudioDescriptionLanguageCodeControl

Choosing followInput will cause the ISO 639 language code of the output to follow the ISO 639 language code of the input. The languageCode will be used when useConfigured is set, or when followInput is selected but there is no ISO 639 language code specified by the input.

codecSettings :: Maybe AudioCodecSettings

Audio codec settings.

audioWatermarkingSettings :: Maybe AudioWatermarkSettings

Settings to configure one or more solutions that insert audio watermarks in the audio encode

streamName :: Maybe Text

Used for MS Smooth and Apple HLS outputs. Indicates the name displayed by the player (eg. English, or Director Commentary).

remixSettings :: Maybe RemixSettings

Settings that control how input audio channels are remixed into the output audio channels.

audioTypeControl :: Maybe AudioDescriptionAudioTypeControl

Determines how audio type is determined. followInput: If the input contains an ISO 639 audioType, then that value is passed through to the output. If the input contains no ISO 639 audioType, the value in Audio Type is included in the output. useConfigured: The value in Audio Type is included in the output. Note that this field and audioType are both ignored if inputType is broadcasterMixedAd.

audioSelectorName :: Text

The name of the AudioSelector used as the source for this AudioDescription.

name :: Text

The name of this AudioDescription. Outputs will use this name to uniquely identify this AudioDescription. Description names should be unique within this Live Event.

newAudioDescription

Arguments

:: Text
$sel:audioSelectorName:AudioDescription'

-> Text
$sel:name:AudioDescription'

-> AudioDescription

Create a value of AudioDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:AudioDescription', audioDescription_languageCode - RFC 5646 language code representing the language of the audio output track. Only used if languageControlMode is useConfigured, or there is no ISO 639 language code specified in the input.

$sel:audioType:AudioDescription', audioDescription_audioType - Applies only if audioTypeControl is useConfigured. The values for audioType are defined in ISO-IEC 13818-1.

$sel:audioNormalizationSettings:AudioDescription', audioDescription_audioNormalizationSettings - Advanced audio normalization settings.

$sel:languageCodeControl:AudioDescription', audioDescription_languageCodeControl - Choosing followInput will cause the ISO 639 language code of the output to follow the ISO 639 language code of the input. The languageCode will be used when useConfigured is set, or when followInput is selected but there is no ISO 639 language code specified by the input.

$sel:codecSettings:AudioDescription', audioDescription_codecSettings - Audio codec settings.

$sel:audioWatermarkingSettings:AudioDescription', audioDescription_audioWatermarkingSettings - Settings to configure one or more solutions that insert audio watermarks in the audio encode

$sel:streamName:AudioDescription', audioDescription_streamName - Used for MS Smooth and Apple HLS outputs. Indicates the name displayed by the player (eg. English, or Director Commentary).

$sel:remixSettings:AudioDescription', audioDescription_remixSettings - Settings that control how input audio channels are remixed into the output audio channels.

$sel:audioTypeControl:AudioDescription', audioDescription_audioTypeControl - Determines how audio type is determined. followInput: If the input contains an ISO 639 audioType, then that value is passed through to the output. If the input contains no ISO 639 audioType, the value in Audio Type is included in the output. useConfigured: The value in Audio Type is included in the output. Note that this field and audioType are both ignored if inputType is broadcasterMixedAd.

$sel:audioSelectorName:AudioDescription', audioDescription_audioSelectorName - The name of the AudioSelector used as the source for this AudioDescription.

$sel:name:AudioDescription', audioDescription_name - The name of this AudioDescription. Outputs will use this name to uniquely identify this AudioDescription. Description names should be unique within this Live Event.

audioDescription_languageCode :: Lens' AudioDescription (Maybe Text)

RFC 5646 language code representing the language of the audio output track. Only used if languageControlMode is useConfigured, or there is no ISO 639 language code specified in the input.

audioDescription_audioType :: Lens' AudioDescription (Maybe AudioType)

Applies only if audioTypeControl is useConfigured. The values for audioType are defined in ISO-IEC 13818-1.

audioDescription_audioNormalizationSettings :: Lens' AudioDescription (Maybe AudioNormalizationSettings)

Advanced audio normalization settings.

audioDescription_languageCodeControl :: Lens' AudioDescription (Maybe AudioDescriptionLanguageCodeControl)

Choosing followInput will cause the ISO 639 language code of the output to follow the ISO 639 language code of the input. The languageCode will be used when useConfigured is set, or when followInput is selected but there is no ISO 639 language code specified by the input.

audioDescription_codecSettings :: Lens' AudioDescription (Maybe AudioCodecSettings)

Audio codec settings.

audioDescription_audioWatermarkingSettings :: Lens' AudioDescription (Maybe AudioWatermarkSettings)

Settings to configure one or more solutions that insert audio watermarks in the audio encode

audioDescription_streamName :: Lens' AudioDescription (Maybe Text)

Used for MS Smooth and Apple HLS outputs. Indicates the name displayed by the player (eg. English, or Director Commentary).

audioDescription_remixSettings :: Lens' AudioDescription (Maybe RemixSettings)

Settings that control how input audio channels are remixed into the output audio channels.

audioDescription_audioTypeControl :: Lens' AudioDescription (Maybe AudioDescriptionAudioTypeControl)

Determines how audio type is determined. followInput: If the input contains an ISO 639 audioType, then that value is passed through to the output. If the input contains no ISO 639 audioType, the value in Audio Type is included in the output. useConfigured: The value in Audio Type is included in the output. Note that this field and audioType are both ignored if inputType is broadcasterMixedAd.

audioDescription_audioSelectorName :: Lens' AudioDescription Text

The name of the AudioSelector used as the source for this AudioDescription.

audioDescription_name :: Lens' AudioDescription Text

The name of this AudioDescription. Outputs will use this name to uniquely identify this AudioDescription. Description names should be unique within this Live Event.