Network.AWS.MediaLive.Types.AudioNormalizationSettings

Description

Documentation

data AudioNormalizationSettings

Audio Normalization Settings

See: newAudioNormalizationSettings smart constructor.

Constructors

AudioNormalizationSettings'

Fields

algorithm :: Maybe AudioNormalizationAlgorithm

Audio normalization algorithm to use. itu17701 conforms to the CALM Act specification, itu17702 conforms to the EBU R-128 specification.

targetLkfs :: Maybe Double

Target LKFS(loudness) to adjust volume to. If no value is entered, a default value will be used according to the chosen algorithm. The CALM Act (1770-1) recommends a target of -24 LKFS. The EBU R-128 specification (1770-2) recommends a target of -23 LKFS.

algorithmControl :: Maybe AudioNormalizationAlgorithmControl

When set to correctAudio the output audio is corrected using the chosen algorithm. If set to measureOnly, the audio will be measured but not adjusted.

newAudioNormalizationSettings :: AudioNormalizationSettings

Create a value of AudioNormalizationSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:algorithm:AudioNormalizationSettings', audioNormalizationSettings_algorithm - Audio normalization algorithm to use. itu17701 conforms to the CALM Act specification, itu17702 conforms to the EBU R-128 specification.

$sel:targetLkfs:AudioNormalizationSettings', audioNormalizationSettings_targetLkfs - Target LKFS(loudness) to adjust volume to. If no value is entered, a default value will be used according to the chosen algorithm. The CALM Act (1770-1) recommends a target of -24 LKFS. The EBU R-128 specification (1770-2) recommends a target of -23 LKFS.

$sel:algorithmControl:AudioNormalizationSettings', audioNormalizationSettings_algorithmControl - When set to correctAudio the output audio is corrected using the chosen algorithm. If set to measureOnly, the audio will be measured but not adjusted.

audioNormalizationSettings_algorithm :: Lens' AudioNormalizationSettings (Maybe AudioNormalizationAlgorithm)

Audio normalization algorithm to use. itu17701 conforms to the CALM Act specification, itu17702 conforms to the EBU R-128 specification.

audioNormalizationSettings_targetLkfs :: Lens' AudioNormalizationSettings (Maybe Double)

Target LKFS(loudness) to adjust volume to. If no value is entered, a default value will be used according to the chosen algorithm. The CALM Act (1770-1) recommends a target of -24 LKFS. The EBU R-128 specification (1770-2) recommends a target of -23 LKFS.

audioNormalizationSettings_algorithmControl :: Lens' AudioNormalizationSettings (Maybe AudioNormalizationAlgorithmControl)

When set to correctAudio the output audio is corrected using the chosen algorithm. If set to measureOnly, the audio will be measured but not adjusted.