Network.AWS.MediaLive.Types.AudioOnlyHlsSettings

Description

Documentation

data AudioOnlyHlsSettings

Audio Only Hls Settings

See: newAudioOnlyHlsSettings smart constructor.

Constructors

AudioOnlyHlsSettings'

Fields

audioGroupId :: Maybe Text

Specifies the group to which the audio Rendition belongs.

audioTrackType :: Maybe AudioOnlyHlsTrackType

Four types of audio-only tracks are supported: Audio-Only Variant Stream The client can play back this audio-only stream instead of video in low-bandwidth scenarios. Represented as an EXT-X-STREAM-INF in the HLS manifest. Alternate Audio, Auto Select, Default Alternate rendition that the client should try to play back by default. Represented as an EXT-X-MEDIA in the HLS manifest with DEFAULT=YES, AUTOSELECT=YES Alternate Audio, Auto Select, Not Default Alternate rendition that the client may try to play back by default. Represented as an EXT-X-MEDIA in the HLS manifest with DEFAULT=NO, AUTOSELECT=YES Alternate Audio, not Auto Select Alternate rendition that the client will not try to play back by default. Represented as an EXT-X-MEDIA in the HLS manifest with DEFAULT=NO, AUTOSELECT=NO

segmentType :: Maybe AudioOnlyHlsSegmentType

Specifies the segment type.

audioOnlyImage :: Maybe InputLocation

Optional. Specifies the .jpg or .png image to use as the cover art for an audio-only output. We recommend a low bit-size file because the image increases the output audio bandwidth. The image is attached to the audio as an ID3 tag, frame type APIC, picture type 0x10, as per the "ID3 tag version 2.4.0 - Native Frames" standard.

newAudioOnlyHlsSettings :: AudioOnlyHlsSettings

Create a value of AudioOnlyHlsSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:audioGroupId:AudioOnlyHlsSettings', audioOnlyHlsSettings_audioGroupId - Specifies the group to which the audio Rendition belongs.

$sel:audioTrackType:AudioOnlyHlsSettings', audioOnlyHlsSettings_audioTrackType - Four types of audio-only tracks are supported: Audio-Only Variant Stream The client can play back this audio-only stream instead of video in low-bandwidth scenarios. Represented as an EXT-X-STREAM-INF in the HLS manifest. Alternate Audio, Auto Select, Default Alternate rendition that the client should try to play back by default. Represented as an EXT-X-MEDIA in the HLS manifest with DEFAULT=YES, AUTOSELECT=YES Alternate Audio, Auto Select, Not Default Alternate rendition that the client may try to play back by default. Represented as an EXT-X-MEDIA in the HLS manifest with DEFAULT=NO, AUTOSELECT=YES Alternate Audio, not Auto Select Alternate rendition that the client will not try to play back by default. Represented as an EXT-X-MEDIA in the HLS manifest with DEFAULT=NO, AUTOSELECT=NO

$sel:segmentType:AudioOnlyHlsSettings', audioOnlyHlsSettings_segmentType - Specifies the segment type.

$sel:audioOnlyImage:AudioOnlyHlsSettings', audioOnlyHlsSettings_audioOnlyImage - Optional. Specifies the .jpg or .png image to use as the cover art for an audio-only output. We recommend a low bit-size file because the image increases the output audio bandwidth. The image is attached to the audio as an ID3 tag, frame type APIC, picture type 0x10, as per the "ID3 tag version 2.4.0 - Native Frames" standard.

audioOnlyHlsSettings_audioGroupId :: Lens' AudioOnlyHlsSettings (Maybe Text)

Specifies the group to which the audio Rendition belongs.

audioOnlyHlsSettings_audioTrackType :: Lens' AudioOnlyHlsSettings (Maybe AudioOnlyHlsTrackType)

Four types of audio-only tracks are supported: Audio-Only Variant Stream The client can play back this audio-only stream instead of video in low-bandwidth scenarios. Represented as an EXT-X-STREAM-INF in the HLS manifest. Alternate Audio, Auto Select, Default Alternate rendition that the client should try to play back by default. Represented as an EXT-X-MEDIA in the HLS manifest with DEFAULT=YES, AUTOSELECT=YES Alternate Audio, Auto Select, Not Default Alternate rendition that the client may try to play back by default. Represented as an EXT-X-MEDIA in the HLS manifest with DEFAULT=NO, AUTOSELECT=YES Alternate Audio, not Auto Select Alternate rendition that the client will not try to play back by default. Represented as an EXT-X-MEDIA in the HLS manifest with DEFAULT=NO, AUTOSELECT=NO

audioOnlyHlsSettings_segmentType :: Lens' AudioOnlyHlsSettings (Maybe AudioOnlyHlsSegmentType)

Specifies the segment type.

audioOnlyHlsSettings_audioOnlyImage :: Lens' AudioOnlyHlsSettings (Maybe InputLocation)

Optional. Specifies the .jpg or .png image to use as the cover art for an audio-only output. We recommend a low bit-size file because the image increases the output audio bandwidth. The image is attached to the audio as an ID3 tag, frame type APIC, picture type 0x10, as per the "ID3 tag version 2.4.0 - Native Frames" standard.