Network.AWS.MediaLive.Types.AutomaticInputFailoverSettings

Description

Documentation

data AutomaticInputFailoverSettings

The settings for Automatic Input Failover.

See: newAutomaticInputFailoverSettings smart constructor.

Constructors

AutomaticInputFailoverSettings'

Fields

failoverConditions :: Maybe [FailoverCondition]

A list of failover conditions. If any of these conditions occur, MediaLive will perform a failover to the other input.

errorClearTimeMsec :: Maybe Natural

This clear time defines the requirement a recovered input must meet to be considered healthy. The input must have no failover conditions for this length of time. Enter a time in milliseconds. This value is particularly important if the input_preference for the failover pair is set to PRIMARY_INPUT_PREFERRED, because after this time, MediaLive will switch back to the primary input.

inputPreference :: Maybe InputPreference

Input preference when deciding which input to make active when a previously failed input has recovered.

secondaryInputId :: Text

The input ID of the secondary input in the automatic input failover pair.

newAutomaticInputFailoverSettings

Arguments

:: Text
$sel:secondaryInputId:AutomaticInputFailoverSettings'

-> AutomaticInputFailoverSettings

Create a value of AutomaticInputFailoverSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failoverConditions:AutomaticInputFailoverSettings', automaticInputFailoverSettings_failoverConditions - A list of failover conditions. If any of these conditions occur, MediaLive will perform a failover to the other input.

$sel:errorClearTimeMsec:AutomaticInputFailoverSettings', automaticInputFailoverSettings_errorClearTimeMsec - This clear time defines the requirement a recovered input must meet to be considered healthy. The input must have no failover conditions for this length of time. Enter a time in milliseconds. This value is particularly important if the input_preference for the failover pair is set to PRIMARY_INPUT_PREFERRED, because after this time, MediaLive will switch back to the primary input.

$sel:inputPreference:AutomaticInputFailoverSettings', automaticInputFailoverSettings_inputPreference - Input preference when deciding which input to make active when a previously failed input has recovered.

$sel:secondaryInputId:AutomaticInputFailoverSettings', automaticInputFailoverSettings_secondaryInputId - The input ID of the secondary input in the automatic input failover pair.

automaticInputFailoverSettings_failoverConditions :: Lens' AutomaticInputFailoverSettings (Maybe [FailoverCondition])

A list of failover conditions. If any of these conditions occur, MediaLive will perform a failover to the other input.

automaticInputFailoverSettings_errorClearTimeMsec :: Lens' AutomaticInputFailoverSettings (Maybe Natural)

This clear time defines the requirement a recovered input must meet to be considered healthy. The input must have no failover conditions for this length of time. Enter a time in milliseconds. This value is particularly important if the input_preference for the failover pair is set to PRIMARY_INPUT_PREFERRED, because after this time, MediaLive will switch back to the primary input.

automaticInputFailoverSettings_inputPreference :: Lens' AutomaticInputFailoverSettings (Maybe InputPreference)

Input preference when deciding which input to make active when a previously failed input has recovered.

automaticInputFailoverSettings_secondaryInputId :: Lens' AutomaticInputFailoverSettings Text

The input ID of the secondary input in the automatic input failover pair.