Network.AWS.MediaLive.Types.AvailBlanking

Description

Documentation

data AvailBlanking

Avail Blanking

See: newAvailBlanking smart constructor.

Constructors

AvailBlanking'

Fields

state :: Maybe AvailBlankingState

When set to enabled, causes video, audio and captions to be blanked when insertion metadata is added.

availBlankingImage :: Maybe InputLocation

Blanking image to be used. Leave empty for solid black. Only bmp and png images are supported.

newAvailBlanking :: AvailBlanking

Create a value of AvailBlanking with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:AvailBlanking', availBlanking_state - When set to enabled, causes video, audio and captions to be blanked when insertion metadata is added.

$sel:availBlankingImage:AvailBlanking', availBlanking_availBlankingImage - Blanking image to be used. Leave empty for solid black. Only bmp and png images are supported.

availBlanking_state :: Lens' AvailBlanking (Maybe AvailBlankingState)

When set to enabled, causes video, audio and captions to be blanked when insertion metadata is added.

availBlanking_availBlankingImage :: Lens' AvailBlanking (Maybe InputLocation)

Blanking image to be used. Leave empty for solid black. Only bmp and png images are supported.