Network.AWS.MediaLive.Types.BlackoutSlate

Description

Documentation

data BlackoutSlate

Blackout Slate

See: newBlackoutSlate smart constructor.

Constructors

BlackoutSlate'

Fields

networkEndBlackoutImage :: Maybe InputLocation

Path to local file to use as Network End Blackout image. Image will be scaled to fill the entire output raster.

state :: Maybe BlackoutSlateState

When set to enabled, causes video, audio and captions to be blanked when indicated by program metadata.

networkEndBlackout :: Maybe BlackoutSlateNetworkEndBlackout

Setting to enabled causes the encoder to blackout the video, audio, and captions, and raise the "Network Blackout Image" slate when an SCTE104/35 Network End Segmentation Descriptor is encountered. The blackout will be lifted when the Network Start Segmentation Descriptor is encountered. The Network End and Network Start descriptors must contain a network ID that matches the value entered in "Network ID".

networkId :: Maybe Text

Provides Network ID that matches EIDR ID format (e.g., "10.XXXX/XXXX-XXXX-XXXX-XXXX-XXXX-C").

blackoutSlateImage :: Maybe InputLocation

Blackout slate image to be used. Leave empty for solid black. Only bmp and png images are supported.

newBlackoutSlate :: BlackoutSlate

Create a value of BlackoutSlate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:networkEndBlackoutImage:BlackoutSlate', blackoutSlate_networkEndBlackoutImage - Path to local file to use as Network End Blackout image. Image will be scaled to fill the entire output raster.

$sel:state:BlackoutSlate', blackoutSlate_state - When set to enabled, causes video, audio and captions to be blanked when indicated by program metadata.

$sel:networkEndBlackout:BlackoutSlate', blackoutSlate_networkEndBlackout - Setting to enabled causes the encoder to blackout the video, audio, and captions, and raise the "Network Blackout Image" slate when an SCTE104/35 Network End Segmentation Descriptor is encountered. The blackout will be lifted when the Network Start Segmentation Descriptor is encountered. The Network End and Network Start descriptors must contain a network ID that matches the value entered in "Network ID".

$sel:networkId:BlackoutSlate', blackoutSlate_networkId - Provides Network ID that matches EIDR ID format (e.g., "10.XXXX/XXXX-XXXX-XXXX-XXXX-XXXX-C").

$sel:blackoutSlateImage:BlackoutSlate', blackoutSlate_blackoutSlateImage - Blackout slate image to be used. Leave empty for solid black. Only bmp and png images are supported.

blackoutSlate_networkEndBlackoutImage :: Lens' BlackoutSlate (Maybe InputLocation)

Path to local file to use as Network End Blackout image. Image will be scaled to fill the entire output raster.

blackoutSlate_state :: Lens' BlackoutSlate (Maybe BlackoutSlateState)

When set to enabled, causes video, audio and captions to be blanked when indicated by program metadata.

blackoutSlate_networkEndBlackout :: Lens' BlackoutSlate (Maybe BlackoutSlateNetworkEndBlackout)

Setting to enabled causes the encoder to blackout the video, audio, and captions, and raise the "Network Blackout Image" slate when an SCTE104/35 Network End Segmentation Descriptor is encountered. The blackout will be lifted when the Network Start Segmentation Descriptor is encountered. The Network End and Network Start descriptors must contain a network ID that matches the value entered in "Network ID".

blackoutSlate_networkId :: Lens' BlackoutSlate (Maybe Text)

Provides Network ID that matches EIDR ID format (e.g., "10.XXXX/XXXX-XXXX-XXXX-XXXX-XXXX-C").

blackoutSlate_blackoutSlateImage :: Lens' BlackoutSlate (Maybe InputLocation)

Blackout slate image to be used. Leave empty for solid black. Only bmp and png images are supported.