Network.AWS.MediaLive.Types.BurnInDestinationSettings

Description

Documentation

data BurnInDestinationSettings

Burn In Destination Settings

See: newBurnInDestinationSettings smart constructor.

Constructors

BurnInDestinationSettings'

Fields

alignment :: Maybe BurnInAlignment

If no explicit xPosition or yPosition is provided, setting alignment to centered will place the captions at the bottom center of the output. Similarly, setting a left alignment will align captions to the bottom left of the output. If x and y positions are given in conjunction with the alignment parameter, the font will be justified (either left or centered) relative to those coordinates. Selecting "smart" justification will left-justify live subtitles and center-justify pre-recorded subtitles. All burn-in and DVB-Sub font settings must match.

shadowOpacity :: Maybe Natural

Specifies the opacity of the shadow. 255 is opaque; 0 is transparent. Leaving this parameter out is equivalent to setting it to 0 (transparent). All burn-in and DVB-Sub font settings must match.

shadowColor :: Maybe BurnInShadowColor

Specifies the color of the shadow cast by the captions. All burn-in and DVB-Sub font settings must match.

outlineColor :: Maybe BurnInOutlineColor

Specifies font outline color. This option is not valid for source captions that are either 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

teletextGridControl :: Maybe BurnInTeletextGridControl

Controls whether a fixed grid size will be used to generate the output subtitles bitmap. Only applicable for Teletext inputs and DVB-Sub/Burn-in outputs.

backgroundOpacity :: Maybe Natural

Specifies the opacity of the background rectangle. 255 is opaque; 0 is transparent. Leaving this parameter out is equivalent to setting it to 0 (transparent). All burn-in and DVB-Sub font settings must match.

xPosition :: Maybe Natural

Specifies the horizontal position of the caption relative to the left side of the output in pixels. A value of 10 would result in the captions starting 10 pixels from the left of the output. If no explicit xPosition is provided, the horizontal caption position will be determined by the alignment parameter. All burn-in and DVB-Sub font settings must match.

fontColor :: Maybe BurnInFontColor

Specifies the color of the burned-in captions. This option is not valid for source captions that are STL, 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

fontSize :: Maybe Text

When set to 'auto' fontSize will scale depending on the size of the output. Giving a positive integer will specify the exact font size in points. All burn-in and DVB-Sub font settings must match.

shadowXOffset :: Maybe Int

Specifies the horizontal offset of the shadow relative to the captions in pixels. A value of -2 would result in a shadow offset 2 pixels to the left. All burn-in and DVB-Sub font settings must match.

font :: Maybe InputLocation

External font file used for caption burn-in. File extension must be 'ttf' or 'tte'. Although the user can select output fonts for many different types of input captions, embedded, STL and teletext sources use a strict grid system. Using external fonts with these caption sources could cause unexpected display of proportional fonts. All burn-in and DVB-Sub font settings must match.

backgroundColor :: Maybe BurnInBackgroundColor

Specifies the color of the rectangle behind the captions. All burn-in and DVB-Sub font settings must match.

yPosition :: Maybe Natural

Specifies the vertical position of the caption relative to the top of the output in pixels. A value of 10 would result in the captions starting 10 pixels from the top of the output. If no explicit yPosition is provided, the caption will be positioned towards the bottom of the output. All burn-in and DVB-Sub font settings must match.

outlineSize :: Maybe Natural

Specifies font outline size in pixels. This option is not valid for source captions that are either 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

fontResolution :: Maybe Natural

Font resolution in DPI (dots per inch); default is 96 dpi. All burn-in and DVB-Sub font settings must match.

shadowYOffset :: Maybe Int

Specifies the vertical offset of the shadow relative to the captions in pixels. A value of -2 would result in a shadow offset 2 pixels above the text. All burn-in and DVB-Sub font settings must match.

fontOpacity :: Maybe Natural

Specifies the opacity of the burned-in captions. 255 is opaque; 0 is transparent. All burn-in and DVB-Sub font settings must match.

newBurnInDestinationSettings :: BurnInDestinationSettings

Create a value of BurnInDestinationSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alignment:BurnInDestinationSettings', burnInDestinationSettings_alignment - If no explicit xPosition or yPosition is provided, setting alignment to centered will place the captions at the bottom center of the output. Similarly, setting a left alignment will align captions to the bottom left of the output. If x and y positions are given in conjunction with the alignment parameter, the font will be justified (either left or centered) relative to those coordinates. Selecting "smart" justification will left-justify live subtitles and center-justify pre-recorded subtitles. All burn-in and DVB-Sub font settings must match.

$sel:shadowOpacity:BurnInDestinationSettings', burnInDestinationSettings_shadowOpacity - Specifies the opacity of the shadow. 255 is opaque; 0 is transparent. Leaving this parameter out is equivalent to setting it to 0 (transparent). All burn-in and DVB-Sub font settings must match.

$sel:shadowColor:BurnInDestinationSettings', burnInDestinationSettings_shadowColor - Specifies the color of the shadow cast by the captions. All burn-in and DVB-Sub font settings must match.

$sel:outlineColor:BurnInDestinationSettings', burnInDestinationSettings_outlineColor - Specifies font outline color. This option is not valid for source captions that are either 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

$sel:teletextGridControl:BurnInDestinationSettings', burnInDestinationSettings_teletextGridControl - Controls whether a fixed grid size will be used to generate the output subtitles bitmap. Only applicable for Teletext inputs and DVB-Sub/Burn-in outputs.

$sel:backgroundOpacity:BurnInDestinationSettings', burnInDestinationSettings_backgroundOpacity - Specifies the opacity of the background rectangle. 255 is opaque; 0 is transparent. Leaving this parameter out is equivalent to setting it to 0 (transparent). All burn-in and DVB-Sub font settings must match.

$sel:xPosition:BurnInDestinationSettings', burnInDestinationSettings_xPosition - Specifies the horizontal position of the caption relative to the left side of the output in pixels. A value of 10 would result in the captions starting 10 pixels from the left of the output. If no explicit xPosition is provided, the horizontal caption position will be determined by the alignment parameter. All burn-in and DVB-Sub font settings must match.

$sel:fontColor:BurnInDestinationSettings', burnInDestinationSettings_fontColor - Specifies the color of the burned-in captions. This option is not valid for source captions that are STL, 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

$sel:fontSize:BurnInDestinationSettings', burnInDestinationSettings_fontSize - When set to 'auto' fontSize will scale depending on the size of the output. Giving a positive integer will specify the exact font size in points. All burn-in and DVB-Sub font settings must match.

$sel:shadowXOffset:BurnInDestinationSettings', burnInDestinationSettings_shadowXOffset - Specifies the horizontal offset of the shadow relative to the captions in pixels. A value of -2 would result in a shadow offset 2 pixels to the left. All burn-in and DVB-Sub font settings must match.

$sel:font:BurnInDestinationSettings', burnInDestinationSettings_font - External font file used for caption burn-in. File extension must be 'ttf' or 'tte'. Although the user can select output fonts for many different types of input captions, embedded, STL and teletext sources use a strict grid system. Using external fonts with these caption sources could cause unexpected display of proportional fonts. All burn-in and DVB-Sub font settings must match.

$sel:backgroundColor:BurnInDestinationSettings', burnInDestinationSettings_backgroundColor - Specifies the color of the rectangle behind the captions. All burn-in and DVB-Sub font settings must match.

$sel:yPosition:BurnInDestinationSettings', burnInDestinationSettings_yPosition - Specifies the vertical position of the caption relative to the top of the output in pixels. A value of 10 would result in the captions starting 10 pixels from the top of the output. If no explicit yPosition is provided, the caption will be positioned towards the bottom of the output. All burn-in and DVB-Sub font settings must match.

$sel:outlineSize:BurnInDestinationSettings', burnInDestinationSettings_outlineSize - Specifies font outline size in pixels. This option is not valid for source captions that are either 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

$sel:fontResolution:BurnInDestinationSettings', burnInDestinationSettings_fontResolution - Font resolution in DPI (dots per inch); default is 96 dpi. All burn-in and DVB-Sub font settings must match.

$sel:shadowYOffset:BurnInDestinationSettings', burnInDestinationSettings_shadowYOffset - Specifies the vertical offset of the shadow relative to the captions in pixels. A value of -2 would result in a shadow offset 2 pixels above the text. All burn-in and DVB-Sub font settings must match.

$sel:fontOpacity:BurnInDestinationSettings', burnInDestinationSettings_fontOpacity - Specifies the opacity of the burned-in captions. 255 is opaque; 0 is transparent. All burn-in and DVB-Sub font settings must match.

burnInDestinationSettings_alignment :: Lens' BurnInDestinationSettings (Maybe BurnInAlignment)

If no explicit xPosition or yPosition is provided, setting alignment to centered will place the captions at the bottom center of the output. Similarly, setting a left alignment will align captions to the bottom left of the output. If x and y positions are given in conjunction with the alignment parameter, the font will be justified (either left or centered) relative to those coordinates. Selecting "smart" justification will left-justify live subtitles and center-justify pre-recorded subtitles. All burn-in and DVB-Sub font settings must match.

burnInDestinationSettings_shadowOpacity :: Lens' BurnInDestinationSettings (Maybe Natural)

Specifies the opacity of the shadow. 255 is opaque; 0 is transparent. Leaving this parameter out is equivalent to setting it to 0 (transparent). All burn-in and DVB-Sub font settings must match.

burnInDestinationSettings_shadowColor :: Lens' BurnInDestinationSettings (Maybe BurnInShadowColor)

Specifies the color of the shadow cast by the captions. All burn-in and DVB-Sub font settings must match.

burnInDestinationSettings_outlineColor :: Lens' BurnInDestinationSettings (Maybe BurnInOutlineColor)

Specifies font outline color. This option is not valid for source captions that are either 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

burnInDestinationSettings_teletextGridControl :: Lens' BurnInDestinationSettings (Maybe BurnInTeletextGridControl)

Controls whether a fixed grid size will be used to generate the output subtitles bitmap. Only applicable for Teletext inputs and DVB-Sub/Burn-in outputs.

burnInDestinationSettings_backgroundOpacity :: Lens' BurnInDestinationSettings (Maybe Natural)

Specifies the opacity of the background rectangle. 255 is opaque; 0 is transparent. Leaving this parameter out is equivalent to setting it to 0 (transparent). All burn-in and DVB-Sub font settings must match.

burnInDestinationSettings_xPosition :: Lens' BurnInDestinationSettings (Maybe Natural)

Specifies the horizontal position of the caption relative to the left side of the output in pixels. A value of 10 would result in the captions starting 10 pixels from the left of the output. If no explicit xPosition is provided, the horizontal caption position will be determined by the alignment parameter. All burn-in and DVB-Sub font settings must match.

burnInDestinationSettings_fontColor :: Lens' BurnInDestinationSettings (Maybe BurnInFontColor)

Specifies the color of the burned-in captions. This option is not valid for source captions that are STL, 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

burnInDestinationSettings_fontSize :: Lens' BurnInDestinationSettings (Maybe Text)

When set to 'auto' fontSize will scale depending on the size of the output. Giving a positive integer will specify the exact font size in points. All burn-in and DVB-Sub font settings must match.

burnInDestinationSettings_shadowXOffset :: Lens' BurnInDestinationSettings (Maybe Int)

Specifies the horizontal offset of the shadow relative to the captions in pixels. A value of -2 would result in a shadow offset 2 pixels to the left. All burn-in and DVB-Sub font settings must match.

burnInDestinationSettings_font :: Lens' BurnInDestinationSettings (Maybe InputLocation)

External font file used for caption burn-in. File extension must be 'ttf' or 'tte'. Although the user can select output fonts for many different types of input captions, embedded, STL and teletext sources use a strict grid system. Using external fonts with these caption sources could cause unexpected display of proportional fonts. All burn-in and DVB-Sub font settings must match.

burnInDestinationSettings_backgroundColor :: Lens' BurnInDestinationSettings (Maybe BurnInBackgroundColor)

Specifies the color of the rectangle behind the captions. All burn-in and DVB-Sub font settings must match.

burnInDestinationSettings_yPosition :: Lens' BurnInDestinationSettings (Maybe Natural)

Specifies the vertical position of the caption relative to the top of the output in pixels. A value of 10 would result in the captions starting 10 pixels from the top of the output. If no explicit yPosition is provided, the caption will be positioned towards the bottom of the output. All burn-in and DVB-Sub font settings must match.

burnInDestinationSettings_outlineSize :: Lens' BurnInDestinationSettings (Maybe Natural)

Specifies font outline size in pixels. This option is not valid for source captions that are either 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

burnInDestinationSettings_fontResolution :: Lens' BurnInDestinationSettings (Maybe Natural)

Font resolution in DPI (dots per inch); default is 96 dpi. All burn-in and DVB-Sub font settings must match.

burnInDestinationSettings_shadowYOffset :: Lens' BurnInDestinationSettings (Maybe Int)

Specifies the vertical offset of the shadow relative to the captions in pixels. A value of -2 would result in a shadow offset 2 pixels above the text. All burn-in and DVB-Sub font settings must match.

burnInDestinationSettings_fontOpacity :: Lens' BurnInDestinationSettings (Maybe Natural)

Specifies the opacity of the burned-in captions. 255 is opaque; 0 is transparent. All burn-in and DVB-Sub font settings must match.