Network.AWS.MediaLive.Types.CaptionDescription

Description

Documentation

data CaptionDescription

Caption Description

See: newCaptionDescription smart constructor.

Constructors

CaptionDescription'

Fields

languageCode :: Maybe Text

ISO 639-2 three-digit code: http://www.loc.gov/standards/iso639-2/

languageDescription :: Maybe Text

Human readable information to indicate captions available for players (eg. English, or Spanish).

destinationSettings :: Maybe CaptionDestinationSettings

Additional settings for captions destination that depend on the destination type.

captionSelectorName :: Text

Specifies which input caption selector to use as a caption source when generating output captions. This field should match a captionSelector name.

name :: Text

Name of the caption description. Used to associate a caption description with an output. Names must be unique within an event.

newCaptionDescription

Arguments

:: Text
$sel:captionSelectorName:CaptionDescription'

-> Text
$sel:name:CaptionDescription'

-> CaptionDescription

Create a value of CaptionDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:CaptionDescription', captionDescription_languageCode - ISO 639-2 three-digit code: http://www.loc.gov/standards/iso639-2/

$sel:languageDescription:CaptionDescription', captionDescription_languageDescription - Human readable information to indicate captions available for players (eg. English, or Spanish).

$sel:destinationSettings:CaptionDescription', captionDescription_destinationSettings - Additional settings for captions destination that depend on the destination type.

$sel:captionSelectorName:CaptionDescription', captionDescription_captionSelectorName - Specifies which input caption selector to use as a caption source when generating output captions. This field should match a captionSelector name.

$sel:name:CaptionDescription', captionDescription_name - Name of the caption description. Used to associate a caption description with an output. Names must be unique within an event.

captionDescription_languageCode :: Lens' CaptionDescription (Maybe Text)

ISO 639-2 three-digit code: http://www.loc.gov/standards/iso639-2/

captionDescription_languageDescription :: Lens' CaptionDescription (Maybe Text)

Human readable information to indicate captions available for players (eg. English, or Spanish).

captionDescription_destinationSettings :: Lens' CaptionDescription (Maybe CaptionDestinationSettings)

Additional settings for captions destination that depend on the destination type.

captionDescription_captionSelectorName :: Lens' CaptionDescription Text

Specifies which input caption selector to use as a caption source when generating output captions. This field should match a captionSelector name.

captionDescription_name :: Lens' CaptionDescription Text

Name of the caption description. Used to associate a caption description with an output. Names must be unique within an event.