Network.AWS.MediaLive.Types.Channel

Description

Documentation

data Channel

Placeholder documentation for Channel

See: newChannel smart constructor.

Constructors

Channel'

Fields

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the role assumed when running the Channel.

encoderSettings :: Maybe EncoderSettings

inputSpecification :: Maybe InputSpecification

Specification of network and file inputs for this channel

arn :: Maybe Text

The unique arn of the channel.

id :: Maybe Text

The unique id of the channel.

pipelinesRunningCount :: Maybe Int

The number of currently healthy pipelines.

channelClass :: Maybe ChannelClass

The class for this channel. STANDARD for a channel with two pipelines or SINGLE_PIPELINE for a channel with one pipeline.

logLevel :: Maybe LogLevel

The log level being written to CloudWatch Logs.

name :: Maybe Text

The name of the channel. (user-mutable)

destinations :: Maybe [OutputDestination]

A list of destinations of the channel. For UDP outputs, there is one destination per output. For other types (HLS, for example), there is one destination per packager.

state :: Maybe ChannelState

inputAttachments :: Maybe [InputAttachment]

List of input attachments for channel.

tags :: Maybe (HashMap Text Text)

A collection of key-value pairs.

pipelineDetails :: Maybe [PipelineDetail]

Runtime details for the pipelines of a running channel.

vpc :: Maybe VpcOutputSettingsDescription

Settings for VPC output

cdiInputSpecification :: Maybe CdiInputSpecification

Specification of CDI inputs for this channel

egressEndpoints :: Maybe [ChannelEgressEndpoint]

The endpoints where outgoing connections initiate from

newChannel :: Channel

Create a value of Channel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleArn:Channel', channel_roleArn - The Amazon Resource Name (ARN) of the role assumed when running the Channel.

$sel:encoderSettings:Channel', channel_encoderSettings - Undocumented member.

$sel:inputSpecification:Channel', channel_inputSpecification - Specification of network and file inputs for this channel

$sel:arn:Channel', channel_arn - The unique arn of the channel.

$sel:id:Channel', channel_id - The unique id of the channel.

$sel:pipelinesRunningCount:Channel', channel_pipelinesRunningCount - The number of currently healthy pipelines.

$sel:channelClass:Channel', channel_channelClass - The class for this channel. STANDARD for a channel with two pipelines or SINGLE_PIPELINE for a channel with one pipeline.

$sel:logLevel:Channel', channel_logLevel - The log level being written to CloudWatch Logs.

$sel:name:Channel', channel_name - The name of the channel. (user-mutable)

$sel:destinations:Channel', channel_destinations - A list of destinations of the channel. For UDP outputs, there is one destination per output. For other types (HLS, for example), there is one destination per packager.

$sel:state:Channel', channel_state - Undocumented member.

$sel:inputAttachments:Channel', channel_inputAttachments - List of input attachments for channel.

$sel:tags:Channel', channel_tags - A collection of key-value pairs.

$sel:pipelineDetails:Channel', channel_pipelineDetails - Runtime details for the pipelines of a running channel.

$sel:vpc:Channel', channel_vpc - Settings for VPC output

$sel:cdiInputSpecification:Channel', channel_cdiInputSpecification - Specification of CDI inputs for this channel

$sel:egressEndpoints:Channel', channel_egressEndpoints - The endpoints where outgoing connections initiate from

channel_roleArn :: Lens' Channel (Maybe Text)

The Amazon Resource Name (ARN) of the role assumed when running the Channel.

channel_encoderSettings :: Lens' Channel (Maybe EncoderSettings)

Undocumented member.

channel_inputSpecification :: Lens' Channel (Maybe InputSpecification)

Specification of network and file inputs for this channel

channel_arn :: Lens' Channel (Maybe Text)

The unique arn of the channel.

channel_id :: Lens' Channel (Maybe Text)

The unique id of the channel.

channel_pipelinesRunningCount :: Lens' Channel (Maybe Int)

The number of currently healthy pipelines.

channel_channelClass :: Lens' Channel (Maybe ChannelClass)

The class for this channel. STANDARD for a channel with two pipelines or SINGLE_PIPELINE for a channel with one pipeline.

channel_logLevel :: Lens' Channel (Maybe LogLevel)

The log level being written to CloudWatch Logs.

channel_name :: Lens' Channel (Maybe Text)

The name of the channel. (user-mutable)

channel_destinations :: Lens' Channel (Maybe [OutputDestination])

A list of destinations of the channel. For UDP outputs, there is one destination per output. For other types (HLS, for example), there is one destination per packager.

channel_state :: Lens' Channel (Maybe ChannelState)

Undocumented member.

channel_inputAttachments :: Lens' Channel (Maybe [InputAttachment])

List of input attachments for channel.

channel_tags :: Lens' Channel (Maybe (HashMap Text Text))

A collection of key-value pairs.

channel_pipelineDetails :: Lens' Channel (Maybe [PipelineDetail])

Runtime details for the pipelines of a running channel.

channel_vpc :: Lens' Channel (Maybe VpcOutputSettingsDescription)

Settings for VPC output

channel_cdiInputSpecification :: Lens' Channel (Maybe CdiInputSpecification)

Specification of CDI inputs for this channel

channel_egressEndpoints :: Lens' Channel (Maybe [ChannelEgressEndpoint])

The endpoints where outgoing connections initiate from