Network.AWS.MediaLive.Types.DvbSdtSettings

Description

Documentation

data DvbSdtSettings

DVB Service Description Table (SDT)

See: newDvbSdtSettings smart constructor.

Constructors

DvbSdtSettings'

Fields

outputSdt :: Maybe DvbSdtOutputSdt

Selects method of inserting SDT information into output stream. The sdtFollow setting copies SDT information from input stream to output stream. The sdtFollowIfPresent setting copies SDT information from input stream to output stream if SDT information is present in the input, otherwise it will fall back on the user-defined values. The sdtManual setting means user will enter the SDT information. The sdtNone setting means output stream will not contain SDT information.

serviceName :: Maybe Text

The service name placed in the serviceDescriptor in the Service Description Table. Maximum length is 256 characters.

serviceProviderName :: Maybe Text

The service provider name placed in the serviceDescriptor in the Service Description Table. Maximum length is 256 characters.

repInterval :: Maybe Natural

The number of milliseconds between instances of this table in the output transport stream.

newDvbSdtSettings :: DvbSdtSettings

Create a value of DvbSdtSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:outputSdt:DvbSdtSettings', dvbSdtSettings_outputSdt - Selects method of inserting SDT information into output stream. The sdtFollow setting copies SDT information from input stream to output stream. The sdtFollowIfPresent setting copies SDT information from input stream to output stream if SDT information is present in the input, otherwise it will fall back on the user-defined values. The sdtManual setting means user will enter the SDT information. The sdtNone setting means output stream will not contain SDT information.

$sel:serviceName:DvbSdtSettings', dvbSdtSettings_serviceName - The service name placed in the serviceDescriptor in the Service Description Table. Maximum length is 256 characters.

$sel:serviceProviderName:DvbSdtSettings', dvbSdtSettings_serviceProviderName - The service provider name placed in the serviceDescriptor in the Service Description Table. Maximum length is 256 characters.

$sel:repInterval:DvbSdtSettings', dvbSdtSettings_repInterval - The number of milliseconds between instances of this table in the output transport stream.

dvbSdtSettings_outputSdt :: Lens' DvbSdtSettings (Maybe DvbSdtOutputSdt)

Selects method of inserting SDT information into output stream. The sdtFollow setting copies SDT information from input stream to output stream. The sdtFollowIfPresent setting copies SDT information from input stream to output stream if SDT information is present in the input, otherwise it will fall back on the user-defined values. The sdtManual setting means user will enter the SDT information. The sdtNone setting means output stream will not contain SDT information.

dvbSdtSettings_serviceName :: Lens' DvbSdtSettings (Maybe Text)

The service name placed in the serviceDescriptor in the Service Description Table. Maximum length is 256 characters.

dvbSdtSettings_serviceProviderName :: Lens' DvbSdtSettings (Maybe Text)

The service provider name placed in the serviceDescriptor in the Service Description Table. Maximum length is 256 characters.

dvbSdtSettings_repInterval :: Lens' DvbSdtSettings (Maybe Natural)

The number of milliseconds between instances of this table in the output transport stream.