Network.AWS.MediaLive.Types.DvbSubDestinationSettings

Description

Documentation

data DvbSubDestinationSettings

Dvb Sub Destination Settings

See: newDvbSubDestinationSettings smart constructor.

Constructors

DvbSubDestinationSettings'

Fields

backgroundOpacity :: Maybe Natural

Specifies the opacity of the background rectangle. 255 is opaque; 0 is transparent. Leaving this parameter blank is equivalent to setting it to 0 (transparent). All burn-in and DVB-Sub font settings must match.

fontOpacity :: Maybe Natural

Specifies the opacity of the burned-in captions. 255 is opaque; 0 is transparent. All burn-in and DVB-Sub font settings must match.

shadowYOffset :: Maybe Int

Specifies the vertical offset of the shadow relative to the captions in pixels. A value of -2 would result in a shadow offset 2 pixels above the text. All burn-in and DVB-Sub font settings must match.

fontResolution :: Maybe Natural

Font resolution in DPI (dots per inch); default is 96 dpi. All burn-in and DVB-Sub font settings must match.

yPosition :: Maybe Natural

Specifies the vertical position of the caption relative to the top of the output in pixels. A value of 10 would result in the captions starting 10 pixels from the top of the output. If no explicit yPosition is provided, the caption will be positioned towards the bottom of the output. This option is not valid for source captions that are STL, 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

backgroundColor :: Maybe DvbSubDestinationBackgroundColor

Specifies the color of the rectangle behind the captions. All burn-in and DVB-Sub font settings must match.

shadowXOffset :: Maybe Int

Specifies the horizontal offset of the shadow relative to the captions in pixels. A value of -2 would result in a shadow offset 2 pixels to the left. All burn-in and DVB-Sub font settings must match.

fontSize :: Maybe Text

When set to auto fontSize will scale depending on the size of the output. Giving a positive integer will specify the exact font size in points. All burn-in and DVB-Sub font settings must match.

xPosition :: Maybe Natural

Specifies the horizontal position of the caption relative to the left side of the output in pixels. A value of 10 would result in the captions starting 10 pixels from the left of the output. If no explicit xPosition is provided, the horizontal caption position will be determined by the alignment parameter. This option is not valid for source captions that are STL, 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

alignment :: Maybe DvbSubDestinationAlignment

If no explicit xPosition or yPosition is provided, setting alignment to centered will place the captions at the bottom center of the output. Similarly, setting a left alignment will align captions to the bottom left of the output. If x and y positions are given in conjunction with the alignment parameter, the font will be justified (either left or centered) relative to those coordinates. Selecting "smart" justification will left-justify live subtitles and center-justify pre-recorded subtitles. This option is not valid for source captions that are STL or 608/embedded. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

shadowOpacity :: Maybe Natural

Specifies the opacity of the shadow. 255 is opaque; 0 is transparent. Leaving this parameter blank is equivalent to setting it to 0 (transparent). All burn-in and DVB-Sub font settings must match.

teletextGridControl :: Maybe DvbSubDestinationTeletextGridControl

Controls whether a fixed grid size will be used to generate the output subtitles bitmap. Only applicable for Teletext inputs and DVB-Sub/Burn-in outputs.

outlineColor :: Maybe DvbSubDestinationOutlineColor

Specifies font outline color. This option is not valid for source captions that are either 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

outlineSize :: Maybe Natural

Specifies font outline size in pixels. This option is not valid for source captions that are either 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

font :: Maybe InputLocation

External font file used for caption burn-in. File extension must be 'ttf' or 'tte'. Although the user can select output fonts for many different types of input captions, embedded, STL and teletext sources use a strict grid system. Using external fonts with these caption sources could cause unexpected display of proportional fonts. All burn-in and DVB-Sub font settings must match.

shadowColor :: Maybe DvbSubDestinationShadowColor

Specifies the color of the shadow cast by the captions. All burn-in and DVB-Sub font settings must match.

fontColor :: Maybe DvbSubDestinationFontColor

Specifies the color of the burned-in captions. This option is not valid for source captions that are STL, 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

newDvbSubDestinationSettings :: DvbSubDestinationSettings

Create a value of DvbSubDestinationSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backgroundOpacity:DvbSubDestinationSettings', dvbSubDestinationSettings_backgroundOpacity - Specifies the opacity of the background rectangle. 255 is opaque; 0 is transparent. Leaving this parameter blank is equivalent to setting it to 0 (transparent). All burn-in and DVB-Sub font settings must match.

$sel:fontOpacity:DvbSubDestinationSettings', dvbSubDestinationSettings_fontOpacity - Specifies the opacity of the burned-in captions. 255 is opaque; 0 is transparent. All burn-in and DVB-Sub font settings must match.

$sel:shadowYOffset:DvbSubDestinationSettings', dvbSubDestinationSettings_shadowYOffset - Specifies the vertical offset of the shadow relative to the captions in pixels. A value of -2 would result in a shadow offset 2 pixels above the text. All burn-in and DVB-Sub font settings must match.

$sel:fontResolution:DvbSubDestinationSettings', dvbSubDestinationSettings_fontResolution - Font resolution in DPI (dots per inch); default is 96 dpi. All burn-in and DVB-Sub font settings must match.

$sel:yPosition:DvbSubDestinationSettings', dvbSubDestinationSettings_yPosition - Specifies the vertical position of the caption relative to the top of the output in pixels. A value of 10 would result in the captions starting 10 pixels from the top of the output. If no explicit yPosition is provided, the caption will be positioned towards the bottom of the output. This option is not valid for source captions that are STL, 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

$sel:backgroundColor:DvbSubDestinationSettings', dvbSubDestinationSettings_backgroundColor - Specifies the color of the rectangle behind the captions. All burn-in and DVB-Sub font settings must match.

$sel:shadowXOffset:DvbSubDestinationSettings', dvbSubDestinationSettings_shadowXOffset - Specifies the horizontal offset of the shadow relative to the captions in pixels. A value of -2 would result in a shadow offset 2 pixels to the left. All burn-in and DVB-Sub font settings must match.

$sel:fontSize:DvbSubDestinationSettings', dvbSubDestinationSettings_fontSize - When set to auto fontSize will scale depending on the size of the output. Giving a positive integer will specify the exact font size in points. All burn-in and DVB-Sub font settings must match.

$sel:xPosition:DvbSubDestinationSettings', dvbSubDestinationSettings_xPosition - Specifies the horizontal position of the caption relative to the left side of the output in pixels. A value of 10 would result in the captions starting 10 pixels from the left of the output. If no explicit xPosition is provided, the horizontal caption position will be determined by the alignment parameter. This option is not valid for source captions that are STL, 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

$sel:alignment:DvbSubDestinationSettings', dvbSubDestinationSettings_alignment - If no explicit xPosition or yPosition is provided, setting alignment to centered will place the captions at the bottom center of the output. Similarly, setting a left alignment will align captions to the bottom left of the output. If x and y positions are given in conjunction with the alignment parameter, the font will be justified (either left or centered) relative to those coordinates. Selecting "smart" justification will left-justify live subtitles and center-justify pre-recorded subtitles. This option is not valid for source captions that are STL or 608/embedded. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

$sel:shadowOpacity:DvbSubDestinationSettings', dvbSubDestinationSettings_shadowOpacity - Specifies the opacity of the shadow. 255 is opaque; 0 is transparent. Leaving this parameter blank is equivalent to setting it to 0 (transparent). All burn-in and DVB-Sub font settings must match.

$sel:teletextGridControl:DvbSubDestinationSettings', dvbSubDestinationSettings_teletextGridControl - Controls whether a fixed grid size will be used to generate the output subtitles bitmap. Only applicable for Teletext inputs and DVB-Sub/Burn-in outputs.

$sel:outlineColor:DvbSubDestinationSettings', dvbSubDestinationSettings_outlineColor - Specifies font outline color. This option is not valid for source captions that are either 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

$sel:outlineSize:DvbSubDestinationSettings', dvbSubDestinationSettings_outlineSize - Specifies font outline size in pixels. This option is not valid for source captions that are either 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

$sel:font:DvbSubDestinationSettings', dvbSubDestinationSettings_font - External font file used for caption burn-in. File extension must be 'ttf' or 'tte'. Although the user can select output fonts for many different types of input captions, embedded, STL and teletext sources use a strict grid system. Using external fonts with these caption sources could cause unexpected display of proportional fonts. All burn-in and DVB-Sub font settings must match.

$sel:shadowColor:DvbSubDestinationSettings', dvbSubDestinationSettings_shadowColor - Specifies the color of the shadow cast by the captions. All burn-in and DVB-Sub font settings must match.

$sel:fontColor:DvbSubDestinationSettings', dvbSubDestinationSettings_fontColor - Specifies the color of the burned-in captions. This option is not valid for source captions that are STL, 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

dvbSubDestinationSettings_backgroundOpacity :: Lens' DvbSubDestinationSettings (Maybe Natural)

Specifies the opacity of the background rectangle. 255 is opaque; 0 is transparent. Leaving this parameter blank is equivalent to setting it to 0 (transparent). All burn-in and DVB-Sub font settings must match.

dvbSubDestinationSettings_fontOpacity :: Lens' DvbSubDestinationSettings (Maybe Natural)

Specifies the opacity of the burned-in captions. 255 is opaque; 0 is transparent. All burn-in and DVB-Sub font settings must match.

dvbSubDestinationSettings_shadowYOffset :: Lens' DvbSubDestinationSettings (Maybe Int)

Specifies the vertical offset of the shadow relative to the captions in pixels. A value of -2 would result in a shadow offset 2 pixels above the text. All burn-in and DVB-Sub font settings must match.

dvbSubDestinationSettings_fontResolution :: Lens' DvbSubDestinationSettings (Maybe Natural)

Font resolution in DPI (dots per inch); default is 96 dpi. All burn-in and DVB-Sub font settings must match.

dvbSubDestinationSettings_yPosition :: Lens' DvbSubDestinationSettings (Maybe Natural)

Specifies the vertical position of the caption relative to the top of the output in pixels. A value of 10 would result in the captions starting 10 pixels from the top of the output. If no explicit yPosition is provided, the caption will be positioned towards the bottom of the output. This option is not valid for source captions that are STL, 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

dvbSubDestinationSettings_backgroundColor :: Lens' DvbSubDestinationSettings (Maybe DvbSubDestinationBackgroundColor)

Specifies the color of the rectangle behind the captions. All burn-in and DVB-Sub font settings must match.

dvbSubDestinationSettings_shadowXOffset :: Lens' DvbSubDestinationSettings (Maybe Int)

Specifies the horizontal offset of the shadow relative to the captions in pixels. A value of -2 would result in a shadow offset 2 pixels to the left. All burn-in and DVB-Sub font settings must match.

dvbSubDestinationSettings_fontSize :: Lens' DvbSubDestinationSettings (Maybe Text)

When set to auto fontSize will scale depending on the size of the output. Giving a positive integer will specify the exact font size in points. All burn-in and DVB-Sub font settings must match.

dvbSubDestinationSettings_xPosition :: Lens' DvbSubDestinationSettings (Maybe Natural)

Specifies the horizontal position of the caption relative to the left side of the output in pixels. A value of 10 would result in the captions starting 10 pixels from the left of the output. If no explicit xPosition is provided, the horizontal caption position will be determined by the alignment parameter. This option is not valid for source captions that are STL, 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

dvbSubDestinationSettings_alignment :: Lens' DvbSubDestinationSettings (Maybe DvbSubDestinationAlignment)

If no explicit xPosition or yPosition is provided, setting alignment to centered will place the captions at the bottom center of the output. Similarly, setting a left alignment will align captions to the bottom left of the output. If x and y positions are given in conjunction with the alignment parameter, the font will be justified (either left or centered) relative to those coordinates. Selecting "smart" justification will left-justify live subtitles and center-justify pre-recorded subtitles. This option is not valid for source captions that are STL or 608/embedded. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

dvbSubDestinationSettings_shadowOpacity :: Lens' DvbSubDestinationSettings (Maybe Natural)

Specifies the opacity of the shadow. 255 is opaque; 0 is transparent. Leaving this parameter blank is equivalent to setting it to 0 (transparent). All burn-in and DVB-Sub font settings must match.

dvbSubDestinationSettings_teletextGridControl :: Lens' DvbSubDestinationSettings (Maybe DvbSubDestinationTeletextGridControl)

Controls whether a fixed grid size will be used to generate the output subtitles bitmap. Only applicable for Teletext inputs and DVB-Sub/Burn-in outputs.

dvbSubDestinationSettings_outlineColor :: Lens' DvbSubDestinationSettings (Maybe DvbSubDestinationOutlineColor)

Specifies font outline color. This option is not valid for source captions that are either 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

dvbSubDestinationSettings_outlineSize :: Lens' DvbSubDestinationSettings (Maybe Natural)

Specifies font outline size in pixels. This option is not valid for source captions that are either 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.

dvbSubDestinationSettings_font :: Lens' DvbSubDestinationSettings (Maybe InputLocation)

External font file used for caption burn-in. File extension must be 'ttf' or 'tte'. Although the user can select output fonts for many different types of input captions, embedded, STL and teletext sources use a strict grid system. Using external fonts with these caption sources could cause unexpected display of proportional fonts. All burn-in and DVB-Sub font settings must match.

dvbSubDestinationSettings_shadowColor :: Lens' DvbSubDestinationSettings (Maybe DvbSubDestinationShadowColor)

Specifies the color of the shadow cast by the captions. All burn-in and DVB-Sub font settings must match.

dvbSubDestinationSettings_fontColor :: Lens' DvbSubDestinationSettings (Maybe DvbSubDestinationFontColor)

Specifies the color of the burned-in captions. This option is not valid for source captions that are STL, 608/embedded or teletext. These source settings are already pre-defined by the caption stream. All burn-in and DVB-Sub font settings must match.