Network.AWS.MediaLive.Types.Eac3Settings

Description

Documentation

data Eac3Settings

Eac3 Settings

See: newEac3Settings smart constructor.

Constructors

Eac3Settings'

Fields

loRoCenterMixLevel :: Maybe Double

Left only/Right only center mix level. Only used for 3/2 coding mode.

ltRtCenterMixLevel :: Maybe Double

Left total/Right total center mix level. Only used for 3/2 coding mode.

dialnorm :: Maybe Natural

Sets the dialnorm for the output. If blank and input audio is Dolby Digital Plus, dialnorm will be passed through.

codingMode :: Maybe Eac3CodingMode

Dolby Digital Plus coding mode. Determines number of channels.

lfeControl :: Maybe Eac3LfeControl

When encoding 3/2 audio, setting to lfe enables the LFE channel

loRoSurroundMixLevel :: Maybe Double

Left only/Right only surround mix level. Only used for 3/2 coding mode.

ltRtSurroundMixLevel :: Maybe Double

Left total/Right total surround mix level. Only used for 3/2 coding mode.

dcFilter :: Maybe Eac3DcFilter

When set to enabled, activates a DC highpass filter for all input channels.

lfeFilter :: Maybe Eac3LfeFilter

When set to enabled, applies a 120Hz lowpass filter to the LFE channel prior to encoding. Only valid with codingMode32 coding mode.

stereoDownmix :: Maybe Eac3StereoDownmix

Stereo downmix preference. Only used for 3/2 coding mode.

drcRf :: Maybe Eac3DrcRf

Sets the profile for heavy Dolby dynamic range compression, ensures that the instantaneous signal peaks do not exceed specified levels.

drcLine :: Maybe Eac3DrcLine

Sets the Dolby dynamic range compression profile.

phaseControl :: Maybe Eac3PhaseControl

When set to shift90Degrees, applies a 90-degree phase shift to the surround channels. Only used for 3/2 coding mode.

surroundExMode :: Maybe Eac3SurroundExMode

When encoding 3/2 audio, sets whether an extra center back surround channel is matrix encoded into the left and right surround channels.

bitstreamMode :: Maybe Eac3BitstreamMode

Specifies the bitstream mode (bsmod) for the emitted E-AC-3 stream. See ATSC A/52-2012 (Annex E) for background on these values.

attenuationControl :: Maybe Eac3AttenuationControl

When set to attenuate3Db, applies a 3 dB attenuation to the surround channels. Only used for 3/2 coding mode.

bitrate :: Maybe Double

Average bitrate in bits/second. Valid bitrates depend on the coding mode.

surroundMode :: Maybe Eac3SurroundMode

When encoding 2/0 audio, sets whether Dolby Surround is matrix encoded into the two channels.

passthroughControl :: Maybe Eac3PassthroughControl

When set to whenPossible, input DD+ audio will be passed through if it is present on the input. This detection is dynamic over the life of the transcode. Inputs that alternate between DD+ and non-DD+ content will have a consistent DD+ output as the system alternates between passthrough and encoding.

metadataControl :: Maybe Eac3MetadataControl

When set to followInput, encoder metadata will be sourced from the DD, DD+, or DolbyE decoder that supplied this audio data. If audio was not supplied from one of these streams, then the static metadata settings will be used.

newEac3Settings :: Eac3Settings

Create a value of Eac3Settings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loRoCenterMixLevel:Eac3Settings', eac3Settings_loRoCenterMixLevel - Left only/Right only center mix level. Only used for 3/2 coding mode.

$sel:ltRtCenterMixLevel:Eac3Settings', eac3Settings_ltRtCenterMixLevel - Left total/Right total center mix level. Only used for 3/2 coding mode.

$sel:dialnorm:Eac3Settings', eac3Settings_dialnorm - Sets the dialnorm for the output. If blank and input audio is Dolby Digital Plus, dialnorm will be passed through.

$sel:codingMode:Eac3Settings', eac3Settings_codingMode - Dolby Digital Plus coding mode. Determines number of channels.

$sel:lfeControl:Eac3Settings', eac3Settings_lfeControl - When encoding 3/2 audio, setting to lfe enables the LFE channel

$sel:loRoSurroundMixLevel:Eac3Settings', eac3Settings_loRoSurroundMixLevel - Left only/Right only surround mix level. Only used for 3/2 coding mode.

$sel:ltRtSurroundMixLevel:Eac3Settings', eac3Settings_ltRtSurroundMixLevel - Left total/Right total surround mix level. Only used for 3/2 coding mode.

$sel:dcFilter:Eac3Settings', eac3Settings_dcFilter - When set to enabled, activates a DC highpass filter for all input channels.

$sel:lfeFilter:Eac3Settings', eac3Settings_lfeFilter - When set to enabled, applies a 120Hz lowpass filter to the LFE channel prior to encoding. Only valid with codingMode32 coding mode.

$sel:stereoDownmix:Eac3Settings', eac3Settings_stereoDownmix - Stereo downmix preference. Only used for 3/2 coding mode.

$sel:drcRf:Eac3Settings', eac3Settings_drcRf - Sets the profile for heavy Dolby dynamic range compression, ensures that the instantaneous signal peaks do not exceed specified levels.

$sel:drcLine:Eac3Settings', eac3Settings_drcLine - Sets the Dolby dynamic range compression profile.

$sel:phaseControl:Eac3Settings', eac3Settings_phaseControl - When set to shift90Degrees, applies a 90-degree phase shift to the surround channels. Only used for 3/2 coding mode.

$sel:surroundExMode:Eac3Settings', eac3Settings_surroundExMode - When encoding 3/2 audio, sets whether an extra center back surround channel is matrix encoded into the left and right surround channels.

$sel:bitstreamMode:Eac3Settings', eac3Settings_bitstreamMode - Specifies the bitstream mode (bsmod) for the emitted E-AC-3 stream. See ATSC A/52-2012 (Annex E) for background on these values.

$sel:attenuationControl:Eac3Settings', eac3Settings_attenuationControl - When set to attenuate3Db, applies a 3 dB attenuation to the surround channels. Only used for 3/2 coding mode.

$sel:bitrate:Eac3Settings', eac3Settings_bitrate - Average bitrate in bits/second. Valid bitrates depend on the coding mode.

$sel:surroundMode:Eac3Settings', eac3Settings_surroundMode - When encoding 2/0 audio, sets whether Dolby Surround is matrix encoded into the two channels.

$sel:passthroughControl:Eac3Settings', eac3Settings_passthroughControl - When set to whenPossible, input DD+ audio will be passed through if it is present on the input. This detection is dynamic over the life of the transcode. Inputs that alternate between DD+ and non-DD+ content will have a consistent DD+ output as the system alternates between passthrough and encoding.

$sel:metadataControl:Eac3Settings', eac3Settings_metadataControl - When set to followInput, encoder metadata will be sourced from the DD, DD+, or DolbyE decoder that supplied this audio data. If audio was not supplied from one of these streams, then the static metadata settings will be used.

eac3Settings_loRoCenterMixLevel :: Lens' Eac3Settings (Maybe Double)

Left only/Right only center mix level. Only used for 3/2 coding mode.

eac3Settings_ltRtCenterMixLevel :: Lens' Eac3Settings (Maybe Double)

Left total/Right total center mix level. Only used for 3/2 coding mode.

eac3Settings_dialnorm :: Lens' Eac3Settings (Maybe Natural)

Sets the dialnorm for the output. If blank and input audio is Dolby Digital Plus, dialnorm will be passed through.

eac3Settings_codingMode :: Lens' Eac3Settings (Maybe Eac3CodingMode)

Dolby Digital Plus coding mode. Determines number of channels.

eac3Settings_lfeControl :: Lens' Eac3Settings (Maybe Eac3LfeControl)

When encoding 3/2 audio, setting to lfe enables the LFE channel

eac3Settings_loRoSurroundMixLevel :: Lens' Eac3Settings (Maybe Double)

Left only/Right only surround mix level. Only used for 3/2 coding mode.

eac3Settings_ltRtSurroundMixLevel :: Lens' Eac3Settings (Maybe Double)

Left total/Right total surround mix level. Only used for 3/2 coding mode.

eac3Settings_dcFilter :: Lens' Eac3Settings (Maybe Eac3DcFilter)

When set to enabled, activates a DC highpass filter for all input channels.

eac3Settings_lfeFilter :: Lens' Eac3Settings (Maybe Eac3LfeFilter)

When set to enabled, applies a 120Hz lowpass filter to the LFE channel prior to encoding. Only valid with codingMode32 coding mode.

eac3Settings_stereoDownmix :: Lens' Eac3Settings (Maybe Eac3StereoDownmix)

Stereo downmix preference. Only used for 3/2 coding mode.

eac3Settings_drcRf :: Lens' Eac3Settings (Maybe Eac3DrcRf)

Sets the profile for heavy Dolby dynamic range compression, ensures that the instantaneous signal peaks do not exceed specified levels.

eac3Settings_drcLine :: Lens' Eac3Settings (Maybe Eac3DrcLine)

Sets the Dolby dynamic range compression profile.

eac3Settings_phaseControl :: Lens' Eac3Settings (Maybe Eac3PhaseControl)

When set to shift90Degrees, applies a 90-degree phase shift to the surround channels. Only used for 3/2 coding mode.

eac3Settings_surroundExMode :: Lens' Eac3Settings (Maybe Eac3SurroundExMode)

When encoding 3/2 audio, sets whether an extra center back surround channel is matrix encoded into the left and right surround channels.

eac3Settings_bitstreamMode :: Lens' Eac3Settings (Maybe Eac3BitstreamMode)

Specifies the bitstream mode (bsmod) for the emitted E-AC-3 stream. See ATSC A/52-2012 (Annex E) for background on these values.

eac3Settings_attenuationControl :: Lens' Eac3Settings (Maybe Eac3AttenuationControl)

When set to attenuate3Db, applies a 3 dB attenuation to the surround channels. Only used for 3/2 coding mode.

eac3Settings_bitrate :: Lens' Eac3Settings (Maybe Double)

Average bitrate in bits/second. Valid bitrates depend on the coding mode.

eac3Settings_surroundMode :: Lens' Eac3Settings (Maybe Eac3SurroundMode)

When encoding 2/0 audio, sets whether Dolby Surround is matrix encoded into the two channels.

eac3Settings_passthroughControl :: Lens' Eac3Settings (Maybe Eac3PassthroughControl)

When set to whenPossible, input DD+ audio will be passed through if it is present on the input. This detection is dynamic over the life of the transcode. Inputs that alternate between DD+ and non-DD+ content will have a consistent DD+ output as the system alternates between passthrough and encoding.

eac3Settings_metadataControl :: Lens' Eac3Settings (Maybe Eac3MetadataControl)

When set to followInput, encoder metadata will be sourced from the DD, DD+, or DolbyE decoder that supplied this audio data. If audio was not supplied from one of these streams, then the static metadata settings will be used.