Network.AWS.MediaLive.Types.EmbeddedSourceSettings

Description

Documentation

data EmbeddedSourceSettings

Embedded Source Settings

See: newEmbeddedSourceSettings smart constructor.

Constructors

EmbeddedSourceSettings'

Fields

scte20Detection :: Maybe EmbeddedScte20Detection

Set to "auto" to handle streams with intermittent and/or non-aligned SCTE-20 and Embedded captions.

convert608To708 :: Maybe EmbeddedConvert608To708

If upconvert, 608 data is both passed through via the "608 compatibility bytes" fields of the 708 wrapper as well as translated into 708. 708 data present in the source content will be discarded.

source608TrackNumber :: Maybe Natural

This field is unused and deprecated.

source608ChannelNumber :: Maybe Natural

Specifies the 608/708 channel number within the video track from which to extract captions. Unused for passthrough.

newEmbeddedSourceSettings :: EmbeddedSourceSettings

Create a value of EmbeddedSourceSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scte20Detection:EmbeddedSourceSettings', embeddedSourceSettings_scte20Detection - Set to "auto" to handle streams with intermittent and/or non-aligned SCTE-20 and Embedded captions.

$sel:convert608To708:EmbeddedSourceSettings', embeddedSourceSettings_convert608To708 - If upconvert, 608 data is both passed through via the "608 compatibility bytes" fields of the 708 wrapper as well as translated into 708. 708 data present in the source content will be discarded.

$sel:source608TrackNumber:EmbeddedSourceSettings', embeddedSourceSettings_source608TrackNumber - This field is unused and deprecated.

$sel:source608ChannelNumber:EmbeddedSourceSettings', embeddedSourceSettings_source608ChannelNumber - Specifies the 608/708 channel number within the video track from which to extract captions. Unused for passthrough.

embeddedSourceSettings_scte20Detection :: Lens' EmbeddedSourceSettings (Maybe EmbeddedScte20Detection)

Set to "auto" to handle streams with intermittent and/or non-aligned SCTE-20 and Embedded captions.

embeddedSourceSettings_convert608To708 :: Lens' EmbeddedSourceSettings (Maybe EmbeddedConvert608To708)

If upconvert, 608 data is both passed through via the "608 compatibility bytes" fields of the 708 wrapper as well as translated into 708. 708 data present in the source content will be discarded.

embeddedSourceSettings_source608TrackNumber :: Lens' EmbeddedSourceSettings (Maybe Natural)

This field is unused and deprecated.

embeddedSourceSettings_source608ChannelNumber :: Lens' EmbeddedSourceSettings (Maybe Natural)

Specifies the 608/708 channel number within the video track from which to extract captions. Unused for passthrough.