Network.AWS.MediaLive.Types.EncoderSettings

Description

Documentation

data EncoderSettings

Encoder Settings

See: newEncoderSettings smart constructor.

Constructors

EncoderSettings'

Fields

motionGraphicsConfiguration :: Maybe MotionGraphicsConfiguration

Settings for motion graphics.

featureActivations :: Maybe FeatureActivations

Feature Activations

globalConfiguration :: Maybe GlobalConfiguration

Configuration settings that apply to the event as a whole.

availConfiguration :: Maybe AvailConfiguration

Event-wide configuration settings for ad avail insertion.

availBlanking :: Maybe AvailBlanking

Settings for ad avail blanking.

nielsenConfiguration :: Maybe NielsenConfiguration

Nielsen configuration settings.

blackoutSlate :: Maybe BlackoutSlate

Settings for blackout slate.

captionDescriptions :: Maybe [CaptionDescription]

Settings for caption decriptions

videoDescriptions :: [VideoDescription]

audioDescriptions :: [AudioDescription]

outputGroups :: [OutputGroup]

timecodeConfig :: TimecodeConfig

Contains settings used to acquire and adjust timecode information from inputs.

newEncoderSettings

Arguments

:: TimecodeConfig
$sel:timecodeConfig:EncoderSettings'

-> EncoderSettings

Create a value of EncoderSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:motionGraphicsConfiguration:EncoderSettings', encoderSettings_motionGraphicsConfiguration - Settings for motion graphics.

$sel:featureActivations:EncoderSettings', encoderSettings_featureActivations - Feature Activations

$sel:globalConfiguration:EncoderSettings', encoderSettings_globalConfiguration - Configuration settings that apply to the event as a whole.

$sel:availConfiguration:EncoderSettings', encoderSettings_availConfiguration - Event-wide configuration settings for ad avail insertion.

$sel:availBlanking:EncoderSettings', encoderSettings_availBlanking - Settings for ad avail blanking.

$sel:nielsenConfiguration:EncoderSettings', encoderSettings_nielsenConfiguration - Nielsen configuration settings.

$sel:blackoutSlate:EncoderSettings', encoderSettings_blackoutSlate - Settings for blackout slate.

$sel:captionDescriptions:EncoderSettings', encoderSettings_captionDescriptions - Settings for caption decriptions

$sel:videoDescriptions:EncoderSettings', encoderSettings_videoDescriptions - Undocumented member.

$sel:audioDescriptions:EncoderSettings', encoderSettings_audioDescriptions - Undocumented member.

$sel:outputGroups:EncoderSettings', encoderSettings_outputGroups - Undocumented member.

$sel:timecodeConfig:EncoderSettings', encoderSettings_timecodeConfig - Contains settings used to acquire and adjust timecode information from inputs.

encoderSettings_motionGraphicsConfiguration :: Lens' EncoderSettings (Maybe MotionGraphicsConfiguration)

Settings for motion graphics.

encoderSettings_featureActivations :: Lens' EncoderSettings (Maybe FeatureActivations)

Feature Activations

encoderSettings_globalConfiguration :: Lens' EncoderSettings (Maybe GlobalConfiguration)

Configuration settings that apply to the event as a whole.

encoderSettings_availConfiguration :: Lens' EncoderSettings (Maybe AvailConfiguration)

Event-wide configuration settings for ad avail insertion.

encoderSettings_availBlanking :: Lens' EncoderSettings (Maybe AvailBlanking)

Settings for ad avail blanking.

encoderSettings_nielsenConfiguration :: Lens' EncoderSettings (Maybe NielsenConfiguration)

Nielsen configuration settings.

encoderSettings_blackoutSlate :: Lens' EncoderSettings (Maybe BlackoutSlate)

Settings for blackout slate.

encoderSettings_captionDescriptions :: Lens' EncoderSettings (Maybe [CaptionDescription])

Settings for caption decriptions

encoderSettings_videoDescriptions :: Lens' EncoderSettings [VideoDescription]

Undocumented member.

encoderSettings_audioDescriptions :: Lens' EncoderSettings [AudioDescription]

Undocumented member.

encoderSettings_outputGroups :: Lens' EncoderSettings [OutputGroup]

Undocumented member.

encoderSettings_timecodeConfig :: Lens' EncoderSettings TimecodeConfig

Contains settings used to acquire and adjust timecode information from inputs.