Network.AWS.MediaLive.Types.FailoverConditionSettings

Description

Documentation

data FailoverConditionSettings

Settings for one failover condition.

See: newFailoverConditionSettings smart constructor.

Constructors

FailoverConditionSettings'

Fields

videoBlackSettings :: Maybe VideoBlackFailoverSettings

MediaLive will perform a failover if content is considered black for the specified period.

inputLossSettings :: Maybe InputLossFailoverSettings

MediaLive will perform a failover if content is not detected in this input for the specified period.

audioSilenceSettings :: Maybe AudioSilenceFailoverSettings

MediaLive will perform a failover if the specified audio selector is silent for the specified period.

newFailoverConditionSettings :: FailoverConditionSettings

Create a value of FailoverConditionSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:videoBlackSettings:FailoverConditionSettings', failoverConditionSettings_videoBlackSettings - MediaLive will perform a failover if content is considered black for the specified period.

$sel:inputLossSettings:FailoverConditionSettings', failoverConditionSettings_inputLossSettings - MediaLive will perform a failover if content is not detected in this input for the specified period.

$sel:audioSilenceSettings:FailoverConditionSettings', failoverConditionSettings_audioSilenceSettings - MediaLive will perform a failover if the specified audio selector is silent for the specified period.

failoverConditionSettings_videoBlackSettings :: Lens' FailoverConditionSettings (Maybe VideoBlackFailoverSettings)

MediaLive will perform a failover if content is considered black for the specified period.

failoverConditionSettings_inputLossSettings :: Lens' FailoverConditionSettings (Maybe InputLossFailoverSettings)

MediaLive will perform a failover if content is not detected in this input for the specified period.

failoverConditionSettings_audioSilenceSettings :: Lens' FailoverConditionSettings (Maybe AudioSilenceFailoverSettings)

MediaLive will perform a failover if the specified audio selector is silent for the specified period.