Network.AWS.MediaLive.Types.FollowModeScheduleActionStartSettings

Description

Documentation

data FollowModeScheduleActionStartSettings

Settings to specify if an action follows another.

See: newFollowModeScheduleActionStartSettings smart constructor.

Constructors

FollowModeScheduleActionStartSettings'

Fields

referenceActionName :: Text

The action name of another action that this one refers to.

followPoint :: FollowPoint

Identifies whether this action starts relative to the start or relative to the end of the reference action.

newFollowModeScheduleActionStartSettings

Arguments

:: Text
$sel:referenceActionName:FollowModeScheduleActionStartSettings'

-> FollowPoint
$sel:followPoint:FollowModeScheduleActionStartSettings'

-> FollowModeScheduleActionStartSettings

Create a value of FollowModeScheduleActionStartSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:referenceActionName:FollowModeScheduleActionStartSettings', followModeScheduleActionStartSettings_referenceActionName - The action name of another action that this one refers to.

$sel:followPoint:FollowModeScheduleActionStartSettings', followModeScheduleActionStartSettings_followPoint - Identifies whether this action starts relative to the start or relative to the end of the reference action.

followModeScheduleActionStartSettings_referenceActionName :: Lens' FollowModeScheduleActionStartSettings Text

The action name of another action that this one refers to.

followModeScheduleActionStartSettings_followPoint :: Lens' FollowModeScheduleActionStartSettings FollowPoint

Identifies whether this action starts relative to the start or relative to the end of the reference action.