Network.AWS.MediaLive.Types.GlobalConfiguration

Description

Documentation

data GlobalConfiguration

Global Configuration

See: newGlobalConfiguration smart constructor.

Constructors

GlobalConfiguration'

Fields

outputLockingMode :: Maybe GlobalConfigurationOutputLockingMode

Indicates how MediaLive pipelines are synchronized. PIPELINE_LOCKING - MediaLive will attempt to synchronize the output of each pipeline to the other. EPOCH_LOCKING - MediaLive will attempt to synchronize the output of each pipeline to the Unix epoch.

inputLossBehavior :: Maybe InputLossBehavior

Settings for system actions when input is lost.

initialAudioGain :: Maybe Int

Value to set the initial audio gain for the Live Event.

supportLowFramerateInputs :: Maybe GlobalConfigurationLowFramerateInputs

Adjusts video input buffer for streams with very low video framerates. This is commonly set to enabled for music channels with less than one video frame per second.

inputEndAction :: Maybe GlobalConfigurationInputEndAction

Indicates the action to take when the current input completes (e.g. end-of-file). When switchAndLoopInputs is configured the encoder will restart at the beginning of the first input. When "none" is configured the encoder will transcode either black, a solid color, or a user specified slate images per the "Input Loss Behavior" configuration until the next input switch occurs (which is controlled through the Channel Schedule API).

outputTimingSource :: Maybe GlobalConfigurationOutputTimingSource

Indicates whether the rate of frames emitted by the Live encoder should be paced by its system clock (which optionally may be locked to another source via NTP) or should be locked to the clock of the source that is providing the input stream.

newGlobalConfiguration :: GlobalConfiguration

Create a value of GlobalConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:outputLockingMode:GlobalConfiguration', globalConfiguration_outputLockingMode - Indicates how MediaLive pipelines are synchronized. PIPELINE_LOCKING - MediaLive will attempt to synchronize the output of each pipeline to the other. EPOCH_LOCKING - MediaLive will attempt to synchronize the output of each pipeline to the Unix epoch.

$sel:inputLossBehavior:GlobalConfiguration', globalConfiguration_inputLossBehavior - Settings for system actions when input is lost.

$sel:initialAudioGain:GlobalConfiguration', globalConfiguration_initialAudioGain - Value to set the initial audio gain for the Live Event.

$sel:supportLowFramerateInputs:GlobalConfiguration', globalConfiguration_supportLowFramerateInputs - Adjusts video input buffer for streams with very low video framerates. This is commonly set to enabled for music channels with less than one video frame per second.

$sel:inputEndAction:GlobalConfiguration', globalConfiguration_inputEndAction - Indicates the action to take when the current input completes (e.g. end-of-file). When switchAndLoopInputs is configured the encoder will restart at the beginning of the first input. When "none" is configured the encoder will transcode either black, a solid color, or a user specified slate images per the "Input Loss Behavior" configuration until the next input switch occurs (which is controlled through the Channel Schedule API).

$sel:outputTimingSource:GlobalConfiguration', globalConfiguration_outputTimingSource - Indicates whether the rate of frames emitted by the Live encoder should be paced by its system clock (which optionally may be locked to another source via NTP) or should be locked to the clock of the source that is providing the input stream.

globalConfiguration_outputLockingMode :: Lens' GlobalConfiguration (Maybe GlobalConfigurationOutputLockingMode)

Indicates how MediaLive pipelines are synchronized. PIPELINE_LOCKING - MediaLive will attempt to synchronize the output of each pipeline to the other. EPOCH_LOCKING - MediaLive will attempt to synchronize the output of each pipeline to the Unix epoch.

globalConfiguration_inputLossBehavior :: Lens' GlobalConfiguration (Maybe InputLossBehavior)

Settings for system actions when input is lost.

globalConfiguration_initialAudioGain :: Lens' GlobalConfiguration (Maybe Int)

Value to set the initial audio gain for the Live Event.

globalConfiguration_supportLowFramerateInputs :: Lens' GlobalConfiguration (Maybe GlobalConfigurationLowFramerateInputs)

Adjusts video input buffer for streams with very low video framerates. This is commonly set to enabled for music channels with less than one video frame per second.

globalConfiguration_inputEndAction :: Lens' GlobalConfiguration (Maybe GlobalConfigurationInputEndAction)

Indicates the action to take when the current input completes (e.g. end-of-file). When switchAndLoopInputs is configured the encoder will restart at the beginning of the first input. When "none" is configured the encoder will transcode either black, a solid color, or a user specified slate images per the "Input Loss Behavior" configuration until the next input switch occurs (which is controlled through the Channel Schedule API).

globalConfiguration_outputTimingSource :: Lens' GlobalConfiguration (Maybe GlobalConfigurationOutputTimingSource)

Indicates whether the rate of frames emitted by the Live encoder should be paced by its system clock (which optionally may be locked to another source via NTP) or should be locked to the clock of the source that is providing the input stream.