Network.AWS.MediaLive.Types.H264Settings

Description

Documentation

data H264Settings

H264 Settings

See: newH264Settings smart constructor.

Constructors

H264Settings'

Fields

temporalAq :: Maybe H264TemporalAq

Temporal makes adjustments within each frame based on temporal variation of content complexity. The value to enter in this field depends on the value in the Adaptive quantization field: If you have set the Adaptive quantization field to Auto, MediaLive ignores any value in this field. MediaLive will determine if temporal AQ is appropriate and will apply the appropriate strength. If you have set the Adaptive quantization field to a strength, you can set this field to Enabled or Disabled. Enabled: MediaLive will apply temporal AQ using the specified strength. Disabled: MediaLive won't apply temporal AQ. If you have set the Adaptive quantization to Disabled, MediaLive ignores any value in this field and doesn't apply temporal AQ.

sceneChangeDetect :: Maybe H264SceneChangeDetect

Scene change detection. - On: inserts I-frames when scene change is detected. - Off: does not force an I-frame when scene change is detected.

scanType :: Maybe H264ScanType

Sets the scan type of the output to progressive or top-field-first interlaced.

timecodeInsertion :: Maybe H264TimecodeInsertionBehavior

Determines how timecodes should be inserted into the video elementary stream. - 'disabled': Do not include timecodes - 'picTimingSei': Pass through picture timing SEI messages from the source specified in Timecode Config

parNumerator :: Maybe Natural

Pixel Aspect Ratio numerator.

afdSignaling :: Maybe AfdSignaling

Indicates that AFD values will be written into the output stream. If afdSignaling is "auto", the system will try to preserve the input AFD value (in cases where multiple AFD values are valid). If set to "fixed", the AFD value will be the value configured in the fixedAfd parameter.

gopSize :: Maybe Double

GOP size (keyframe interval) in units of either frames or seconds per gopSizeUnits. If gopSizeUnits is frames, gopSize must be an integer and must be greater than or equal to 1. If gopSizeUnits is seconds, gopSize must be greater than 0, but need not be an integer.

gopSizeUnits :: Maybe H264GopSizeUnits

Indicates if the gopSize is specified in frames or seconds. If seconds the system will convert the gopSize into a frame count at run time.

subgopLength :: Maybe H264SubGopLength

If set to fixed, use gopNumBFrames B-frames per sub-GOP. If set to dynamic, optimize the number of B-frames used for each sub-GOP to improve visual quality.

qualityLevel :: Maybe H264QualityLevel

Leave as STANDARD_QUALITY or choose a different value (which might result in additional costs to run the channel). - ENHANCED_QUALITY: Produces a slightly better video quality without an increase in the bitrate. Has an effect only when the Rate control mode is QVBR or CBR. If this channel is in a MediaLive multiplex, the value must be ENHANCED_QUALITY. - STANDARD_QUALITY: Valid for any Rate control mode.

slices :: Maybe Natural

Number of slices per picture. Must be less than or equal to the number of macroblock rows for progressive pictures, and less than or equal to half the number of macroblock rows for interlaced pictures. This field is optional; when no value is specified the encoder will choose the number of slices based on encode resolution.

profile :: Maybe H264Profile

H.264 Profile.

rateControlMode :: Maybe H264RateControlMode

Rate control mode. QVBR: Quality will match the specified quality level except when it is constrained by the maximum bitrate. Recommended if you or your viewers pay for bandwidth. VBR: Quality and bitrate vary, depending on the video complexity. Recommended instead of QVBR if you want to maintain a specific average bitrate over the duration of the channel. CBR: Quality varies, depending on the video complexity. Recommended only if you distribute your assets to devices that cannot handle variable bitrates. Multiplex: This rate control mode is only supported (and is required) when the video is being delivered to a MediaLive Multiplex in which case the rate control configuration is controlled by the properties within the Multiplex Program.

minIInterval :: Maybe Natural

Only meaningful if sceneChangeDetect is set to enabled. Defaults to 5 if multiplex rate control is used. Enforces separation between repeated (cadence) I-frames and I-frames inserted by Scene Change Detection. If a scene change I-frame is within I-interval frames of a cadence I-frame, the GOP is shrunk and/or stretched to the scene change I-frame. GOP stretch requires enabling lookahead as well as setting I-interval. The normal cadence resumes for the next GOP. Note: Maximum GOP stretch = GOP size + Min-I-interval - 1

qvbrQualityLevel :: Maybe Natural

Controls the target quality for the video encode. Applies only when the rate control mode is QVBR. You can set a target quality or you can let MediaLive determine the best quality. To set a target quality, enter values in the QVBR quality level field and the Max bitrate field. Enter values that suit your most important viewing devices. Recommended values are: - Primary screen: Quality level: 8 to 10. Max bitrate: 4M - PC or tablet: Quality level: 7. Max bitrate: 1.5M to 3M - Smartphone: Quality level: 6. Max bitrate: 1M to 1.5M To let MediaLive decide, leave the QVBR quality level field empty, and in Max bitrate enter the maximum rate you want in the video. For more information, see the section called "Video - rate control mode" in the MediaLive user guide

colorSpaceSettings :: Maybe H264ColorSpaceSettings

Color Space settings

parControl :: Maybe H264ParControl

This field indicates how the output pixel aspect ratio is specified. If "specified" is selected then the output video pixel aspect ratio is determined by parNumerator and parDenominator, else if "initializeFromSource" is selected then the output pixsel aspect ratio will be set equal to the input video pixel aspect ratio of the first input.

flickerAq :: Maybe H264FlickerAq

Flicker AQ makes adjustments within each frame to reduce flicker or 'pop' on I-frames. The value to enter in this field depends on the value in the Adaptive quantization field: If you have set the Adaptive quantization field to Auto, MediaLive ignores any value in this field. MediaLive will determine if flicker AQ is appropriate and will apply the appropriate strength. If you have set the Adaptive quantization field to a strength, you can set this field to Enabled or Disabled. Enabled: MediaLive will apply flicker AQ using the specified strength. Disabled: MediaLive won't apply flicker AQ. If you have set the Adaptive quantization to Disabled, MediaLive ignores any value in this field and doesn't apply flicker AQ.

bufSize :: Maybe Natural

Size of buffer (HRD buffer model) in bits.

spatialAq :: Maybe H264SpatialAq

Spatial AQ makes adjustments within each frame based on spatial variation of content complexity. The value to enter in this field depends on the value in the Adaptive quantization field: If you have set the Adaptive quantization field to Auto, MediaLive ignores any value in this field. MediaLive will determine if spatial AQ is appropriate and will apply the appropriate strength. If you have set the Adaptive quantization field to a strength, you can set this field to Enabled or Disabled. Enabled: MediaLive will apply spatial AQ using the specified strength. Disabled: MediaLive won't apply spatial AQ. If you have set the Adaptive quantization to Disabled, MediaLive ignores any value in this field and doesn't apply spatial AQ.

gopNumBFrames :: Maybe Natural

Number of B-frames between reference frames.

fixedAfd :: Maybe FixedAfd

Four bit AFD value to write on all frames of video in the output stream. Only valid when afdSignaling is set to 'Fixed'.

softness :: Maybe Natural

Softness. Selects quantizer matrix, larger values reduce high-frequency content in the encoded image. If not set to zero, must be greater than 15.

filterSettings :: Maybe H264FilterSettings

Optional filters that you can apply to an encode.

bitrate :: Maybe Natural

Average bitrate in bits/second. Required when the rate control mode is VBR or CBR. Not used for QVBR. In an MS Smooth output group, each output must have a unique value when its bitrate is rounded down to the nearest multiple of 1000.

framerateDenominator :: Maybe Natural

Framerate denominator.

forceFieldPictures :: Maybe H264ForceFieldPictures

This setting applies only when scan type is "interlaced." It controls whether coding is performed on a field basis or on a frame basis. (When the video is progressive, the coding is always performed on a frame basis.) enabled: Force MediaLive to code on a field basis, so that odd and even sets of fields are coded separately. disabled: Code the two sets of fields separately (on a field basis) or together (on a frame basis using PAFF), depending on what is most appropriate for the content.

entropyEncoding :: Maybe H264EntropyEncoding

Entropy encoding mode. Use cabac (must be in Main or High profile) or cavlc.

framerateControl :: Maybe H264FramerateControl

This field indicates how the output video frame rate is specified. If "specified" is selected then the output video frame rate is determined by framerateNumerator and framerateDenominator, else if "initializeFromSource" is selected then the output video frame rate will be set equal to the input video frame rate of the first input.

colorMetadata :: Maybe H264ColorMetadata

Includes colorspace metadata in the output.

lookAheadRateControl :: Maybe H264LookAheadRateControl

Amount of lookahead. A value of low can decrease latency and memory usage, while high can produce better quality for certain content.

adaptiveQuantization :: Maybe H264AdaptiveQuantization

Enables or disables adaptive quantization, which is a technique MediaLive can apply to video on a frame-by-frame basis to produce more compression without losing quality. There are three types of adaptive quantization: flicker, spatial, and temporal. Set the field in one of these ways: Set to Auto. Recommended. For each type of AQ, MediaLive will determine if AQ is needed, and if so, the appropriate strength. Set a strength (a value other than Auto or Disable). This strength will apply to any of the AQ fields that you choose to enable. Set to Disabled to disable all types of adaptive quantization.

framerateNumerator :: Maybe Natural

Framerate numerator - framerate is a fraction, e.g. 24000 / 1001 = 23.976 fps.

level :: Maybe H264Level

H.264 Level.

gopBReference :: Maybe H264GopBReference

Documentation update needed

maxBitrate :: Maybe Natural

For QVBR: See the tooltip for Quality level For VBR: Set the maximum bitrate in order to accommodate expected spikes in the complexity of the video.

syntax :: Maybe H264Syntax

Produces a bitstream compliant with SMPTE RP-2027.

bufFillPct :: Maybe Natural

Percentage of the buffer that should initially be filled (HRD buffer model).

gopClosedCadence :: Maybe Natural

Frequency of closed GOPs. In streaming applications, it is recommended that this be set to 1 so a decoder joining mid-stream will receive an IDR frame as quickly as possible. Setting this value to 0 will break output segmenting.

numRefFrames :: Maybe Natural

Number of reference frames to use. The encoder may use more than requested if using B-frames and/or interlaced encoding.

parDenominator :: Maybe Natural

Pixel Aspect Ratio denominator.

newH264Settings :: H264Settings

Create a value of H264Settings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:temporalAq:H264Settings', h264Settings_temporalAq - Temporal makes adjustments within each frame based on temporal variation of content complexity. The value to enter in this field depends on the value in the Adaptive quantization field: If you have set the Adaptive quantization field to Auto, MediaLive ignores any value in this field. MediaLive will determine if temporal AQ is appropriate and will apply the appropriate strength. If you have set the Adaptive quantization field to a strength, you can set this field to Enabled or Disabled. Enabled: MediaLive will apply temporal AQ using the specified strength. Disabled: MediaLive won't apply temporal AQ. If you have set the Adaptive quantization to Disabled, MediaLive ignores any value in this field and doesn't apply temporal AQ.

$sel:sceneChangeDetect:H264Settings', h264Settings_sceneChangeDetect - Scene change detection. - On: inserts I-frames when scene change is detected. - Off: does not force an I-frame when scene change is detected.

$sel:scanType:H264Settings', h264Settings_scanType - Sets the scan type of the output to progressive or top-field-first interlaced.

$sel:timecodeInsertion:H264Settings', h264Settings_timecodeInsertion - Determines how timecodes should be inserted into the video elementary stream. - 'disabled': Do not include timecodes - 'picTimingSei': Pass through picture timing SEI messages from the source specified in Timecode Config

$sel:parNumerator:H264Settings', h264Settings_parNumerator - Pixel Aspect Ratio numerator.

$sel:afdSignaling:H264Settings', h264Settings_afdSignaling - Indicates that AFD values will be written into the output stream. If afdSignaling is "auto", the system will try to preserve the input AFD value (in cases where multiple AFD values are valid). If set to "fixed", the AFD value will be the value configured in the fixedAfd parameter.

$sel:gopSize:H264Settings', h264Settings_gopSize - GOP size (keyframe interval) in units of either frames or seconds per gopSizeUnits. If gopSizeUnits is frames, gopSize must be an integer and must be greater than or equal to 1. If gopSizeUnits is seconds, gopSize must be greater than 0, but need not be an integer.

$sel:gopSizeUnits:H264Settings', h264Settings_gopSizeUnits - Indicates if the gopSize is specified in frames or seconds. If seconds the system will convert the gopSize into a frame count at run time.

$sel:subgopLength:H264Settings', h264Settings_subgopLength - If set to fixed, use gopNumBFrames B-frames per sub-GOP. If set to dynamic, optimize the number of B-frames used for each sub-GOP to improve visual quality.

$sel:qualityLevel:H264Settings', h264Settings_qualityLevel - Leave as STANDARD_QUALITY or choose a different value (which might result in additional costs to run the channel). - ENHANCED_QUALITY: Produces a slightly better video quality without an increase in the bitrate. Has an effect only when the Rate control mode is QVBR or CBR. If this channel is in a MediaLive multiplex, the value must be ENHANCED_QUALITY. - STANDARD_QUALITY: Valid for any Rate control mode.

$sel:slices:H264Settings', h264Settings_slices - Number of slices per picture. Must be less than or equal to the number of macroblock rows for progressive pictures, and less than or equal to half the number of macroblock rows for interlaced pictures. This field is optional; when no value is specified the encoder will choose the number of slices based on encode resolution.

$sel:profile:H264Settings', h264Settings_profile - H.264 Profile.

$sel:rateControlMode:H264Settings', h264Settings_rateControlMode - Rate control mode. QVBR: Quality will match the specified quality level except when it is constrained by the maximum bitrate. Recommended if you or your viewers pay for bandwidth. VBR: Quality and bitrate vary, depending on the video complexity. Recommended instead of QVBR if you want to maintain a specific average bitrate over the duration of the channel. CBR: Quality varies, depending on the video complexity. Recommended only if you distribute your assets to devices that cannot handle variable bitrates. Multiplex: This rate control mode is only supported (and is required) when the video is being delivered to a MediaLive Multiplex in which case the rate control configuration is controlled by the properties within the Multiplex Program.

$sel:minIInterval:H264Settings', h264Settings_minIInterval - Only meaningful if sceneChangeDetect is set to enabled. Defaults to 5 if multiplex rate control is used. Enforces separation between repeated (cadence) I-frames and I-frames inserted by Scene Change Detection. If a scene change I-frame is within I-interval frames of a cadence I-frame, the GOP is shrunk and/or stretched to the scene change I-frame. GOP stretch requires enabling lookahead as well as setting I-interval. The normal cadence resumes for the next GOP. Note: Maximum GOP stretch = GOP size + Min-I-interval - 1

$sel:qvbrQualityLevel:H264Settings', h264Settings_qvbrQualityLevel - Controls the target quality for the video encode. Applies only when the rate control mode is QVBR. You can set a target quality or you can let MediaLive determine the best quality. To set a target quality, enter values in the QVBR quality level field and the Max bitrate field. Enter values that suit your most important viewing devices. Recommended values are: - Primary screen: Quality level: 8 to 10. Max bitrate: 4M - PC or tablet: Quality level: 7. Max bitrate: 1.5M to 3M - Smartphone: Quality level: 6. Max bitrate: 1M to 1.5M To let MediaLive decide, leave the QVBR quality level field empty, and in Max bitrate enter the maximum rate you want in the video. For more information, see the section called "Video - rate control mode" in the MediaLive user guide

$sel:colorSpaceSettings:H264Settings', h264Settings_colorSpaceSettings - Color Space settings

$sel:parControl:H264Settings', h264Settings_parControl - This field indicates how the output pixel aspect ratio is specified. If "specified" is selected then the output video pixel aspect ratio is determined by parNumerator and parDenominator, else if "initializeFromSource" is selected then the output pixsel aspect ratio will be set equal to the input video pixel aspect ratio of the first input.

$sel:flickerAq:H264Settings', h264Settings_flickerAq - Flicker AQ makes adjustments within each frame to reduce flicker or 'pop' on I-frames. The value to enter in this field depends on the value in the Adaptive quantization field: If you have set the Adaptive quantization field to Auto, MediaLive ignores any value in this field. MediaLive will determine if flicker AQ is appropriate and will apply the appropriate strength. If you have set the Adaptive quantization field to a strength, you can set this field to Enabled or Disabled. Enabled: MediaLive will apply flicker AQ using the specified strength. Disabled: MediaLive won't apply flicker AQ. If you have set the Adaptive quantization to Disabled, MediaLive ignores any value in this field and doesn't apply flicker AQ.

$sel:bufSize:H264Settings', h264Settings_bufSize - Size of buffer (HRD buffer model) in bits.

$sel:spatialAq:H264Settings', h264Settings_spatialAq - Spatial AQ makes adjustments within each frame based on spatial variation of content complexity. The value to enter in this field depends on the value in the Adaptive quantization field: If you have set the Adaptive quantization field to Auto, MediaLive ignores any value in this field. MediaLive will determine if spatial AQ is appropriate and will apply the appropriate strength. If you have set the Adaptive quantization field to a strength, you can set this field to Enabled or Disabled. Enabled: MediaLive will apply spatial AQ using the specified strength. Disabled: MediaLive won't apply spatial AQ. If you have set the Adaptive quantization to Disabled, MediaLive ignores any value in this field and doesn't apply spatial AQ.

$sel:gopNumBFrames:H264Settings', h264Settings_gopNumBFrames - Number of B-frames between reference frames.

$sel:fixedAfd:H264Settings', h264Settings_fixedAfd - Four bit AFD value to write on all frames of video in the output stream. Only valid when afdSignaling is set to 'Fixed'.

$sel:softness:H264Settings', h264Settings_softness - Softness. Selects quantizer matrix, larger values reduce high-frequency content in the encoded image. If not set to zero, must be greater than 15.

$sel:filterSettings:H264Settings', h264Settings_filterSettings - Optional filters that you can apply to an encode.

$sel:bitrate:H264Settings', h264Settings_bitrate - Average bitrate in bits/second. Required when the rate control mode is VBR or CBR. Not used for QVBR. In an MS Smooth output group, each output must have a unique value when its bitrate is rounded down to the nearest multiple of 1000.

$sel:framerateDenominator:H264Settings', h264Settings_framerateDenominator - Framerate denominator.

$sel:forceFieldPictures:H264Settings', h264Settings_forceFieldPictures - This setting applies only when scan type is "interlaced." It controls whether coding is performed on a field basis or on a frame basis. (When the video is progressive, the coding is always performed on a frame basis.) enabled: Force MediaLive to code on a field basis, so that odd and even sets of fields are coded separately. disabled: Code the two sets of fields separately (on a field basis) or together (on a frame basis using PAFF), depending on what is most appropriate for the content.

$sel:entropyEncoding:H264Settings', h264Settings_entropyEncoding - Entropy encoding mode. Use cabac (must be in Main or High profile) or cavlc.

$sel:framerateControl:H264Settings', h264Settings_framerateControl - This field indicates how the output video frame rate is specified. If "specified" is selected then the output video frame rate is determined by framerateNumerator and framerateDenominator, else if "initializeFromSource" is selected then the output video frame rate will be set equal to the input video frame rate of the first input.

$sel:colorMetadata:H264Settings', h264Settings_colorMetadata - Includes colorspace metadata in the output.

$sel:lookAheadRateControl:H264Settings', h264Settings_lookAheadRateControl - Amount of lookahead. A value of low can decrease latency and memory usage, while high can produce better quality for certain content.

$sel:adaptiveQuantization:H264Settings', h264Settings_adaptiveQuantization - Enables or disables adaptive quantization, which is a technique MediaLive can apply to video on a frame-by-frame basis to produce more compression without losing quality. There are three types of adaptive quantization: flicker, spatial, and temporal. Set the field in one of these ways: Set to Auto. Recommended. For each type of AQ, MediaLive will determine if AQ is needed, and if so, the appropriate strength. Set a strength (a value other than Auto or Disable). This strength will apply to any of the AQ fields that you choose to enable. Set to Disabled to disable all types of adaptive quantization.

$sel:framerateNumerator:H264Settings', h264Settings_framerateNumerator - Framerate numerator - framerate is a fraction, e.g. 24000 / 1001 = 23.976 fps.

$sel:level:H264Settings', h264Settings_level - H.264 Level.

$sel:gopBReference:H264Settings', h264Settings_gopBReference - Documentation update needed

$sel:maxBitrate:H264Settings', h264Settings_maxBitrate - For QVBR: See the tooltip for Quality level For VBR: Set the maximum bitrate in order to accommodate expected spikes in the complexity of the video.

$sel:syntax:H264Settings', h264Settings_syntax - Produces a bitstream compliant with SMPTE RP-2027.

$sel:bufFillPct:H264Settings', h264Settings_bufFillPct - Percentage of the buffer that should initially be filled (HRD buffer model).

$sel:gopClosedCadence:H264Settings', h264Settings_gopClosedCadence - Frequency of closed GOPs. In streaming applications, it is recommended that this be set to 1 so a decoder joining mid-stream will receive an IDR frame as quickly as possible. Setting this value to 0 will break output segmenting.

$sel:numRefFrames:H264Settings', h264Settings_numRefFrames - Number of reference frames to use. The encoder may use more than requested if using B-frames and/or interlaced encoding.

$sel:parDenominator:H264Settings', h264Settings_parDenominator - Pixel Aspect Ratio denominator.

h264Settings_temporalAq :: Lens' H264Settings (Maybe H264TemporalAq)

Temporal makes adjustments within each frame based on temporal variation of content complexity. The value to enter in this field depends on the value in the Adaptive quantization field: If you have set the Adaptive quantization field to Auto, MediaLive ignores any value in this field. MediaLive will determine if temporal AQ is appropriate and will apply the appropriate strength. If you have set the Adaptive quantization field to a strength, you can set this field to Enabled or Disabled. Enabled: MediaLive will apply temporal AQ using the specified strength. Disabled: MediaLive won't apply temporal AQ. If you have set the Adaptive quantization to Disabled, MediaLive ignores any value in this field and doesn't apply temporal AQ.

h264Settings_sceneChangeDetect :: Lens' H264Settings (Maybe H264SceneChangeDetect)

Scene change detection. - On: inserts I-frames when scene change is detected. - Off: does not force an I-frame when scene change is detected.

h264Settings_scanType :: Lens' H264Settings (Maybe H264ScanType)

Sets the scan type of the output to progressive or top-field-first interlaced.

h264Settings_timecodeInsertion :: Lens' H264Settings (Maybe H264TimecodeInsertionBehavior)

Determines how timecodes should be inserted into the video elementary stream. - 'disabled': Do not include timecodes - 'picTimingSei': Pass through picture timing SEI messages from the source specified in Timecode Config

h264Settings_parNumerator :: Lens' H264Settings (Maybe Natural)

Pixel Aspect Ratio numerator.

h264Settings_afdSignaling :: Lens' H264Settings (Maybe AfdSignaling)

Indicates that AFD values will be written into the output stream. If afdSignaling is "auto", the system will try to preserve the input AFD value (in cases where multiple AFD values are valid). If set to "fixed", the AFD value will be the value configured in the fixedAfd parameter.

h264Settings_gopSize :: Lens' H264Settings (Maybe Double)

GOP size (keyframe interval) in units of either frames or seconds per gopSizeUnits. If gopSizeUnits is frames, gopSize must be an integer and must be greater than or equal to 1. If gopSizeUnits is seconds, gopSize must be greater than 0, but need not be an integer.

h264Settings_gopSizeUnits :: Lens' H264Settings (Maybe H264GopSizeUnits)

Indicates if the gopSize is specified in frames or seconds. If seconds the system will convert the gopSize into a frame count at run time.

h264Settings_subgopLength :: Lens' H264Settings (Maybe H264SubGopLength)

If set to fixed, use gopNumBFrames B-frames per sub-GOP. If set to dynamic, optimize the number of B-frames used for each sub-GOP to improve visual quality.

h264Settings_qualityLevel :: Lens' H264Settings (Maybe H264QualityLevel)

Leave as STANDARD_QUALITY or choose a different value (which might result in additional costs to run the channel). - ENHANCED_QUALITY: Produces a slightly better video quality without an increase in the bitrate. Has an effect only when the Rate control mode is QVBR or CBR. If this channel is in a MediaLive multiplex, the value must be ENHANCED_QUALITY. - STANDARD_QUALITY: Valid for any Rate control mode.

h264Settings_slices :: Lens' H264Settings (Maybe Natural)

Number of slices per picture. Must be less than or equal to the number of macroblock rows for progressive pictures, and less than or equal to half the number of macroblock rows for interlaced pictures. This field is optional; when no value is specified the encoder will choose the number of slices based on encode resolution.

h264Settings_profile :: Lens' H264Settings (Maybe H264Profile)

H.264 Profile.

h264Settings_rateControlMode :: Lens' H264Settings (Maybe H264RateControlMode)

Rate control mode. QVBR: Quality will match the specified quality level except when it is constrained by the maximum bitrate. Recommended if you or your viewers pay for bandwidth. VBR: Quality and bitrate vary, depending on the video complexity. Recommended instead of QVBR if you want to maintain a specific average bitrate over the duration of the channel. CBR: Quality varies, depending on the video complexity. Recommended only if you distribute your assets to devices that cannot handle variable bitrates. Multiplex: This rate control mode is only supported (and is required) when the video is being delivered to a MediaLive Multiplex in which case the rate control configuration is controlled by the properties within the Multiplex Program.

h264Settings_minIInterval :: Lens' H264Settings (Maybe Natural)

Only meaningful if sceneChangeDetect is set to enabled. Defaults to 5 if multiplex rate control is used. Enforces separation between repeated (cadence) I-frames and I-frames inserted by Scene Change Detection. If a scene change I-frame is within I-interval frames of a cadence I-frame, the GOP is shrunk and/or stretched to the scene change I-frame. GOP stretch requires enabling lookahead as well as setting I-interval. The normal cadence resumes for the next GOP. Note: Maximum GOP stretch = GOP size + Min-I-interval - 1

h264Settings_qvbrQualityLevel :: Lens' H264Settings (Maybe Natural)

Controls the target quality for the video encode. Applies only when the rate control mode is QVBR. You can set a target quality or you can let MediaLive determine the best quality. To set a target quality, enter values in the QVBR quality level field and the Max bitrate field. Enter values that suit your most important viewing devices. Recommended values are: - Primary screen: Quality level: 8 to 10. Max bitrate: 4M - PC or tablet: Quality level: 7. Max bitrate: 1.5M to 3M - Smartphone: Quality level: 6. Max bitrate: 1M to 1.5M To let MediaLive decide, leave the QVBR quality level field empty, and in Max bitrate enter the maximum rate you want in the video. For more information, see the section called "Video - rate control mode" in the MediaLive user guide

h264Settings_colorSpaceSettings :: Lens' H264Settings (Maybe H264ColorSpaceSettings)

Color Space settings

h264Settings_parControl :: Lens' H264Settings (Maybe H264ParControl)

This field indicates how the output pixel aspect ratio is specified. If "specified" is selected then the output video pixel aspect ratio is determined by parNumerator and parDenominator, else if "initializeFromSource" is selected then the output pixsel aspect ratio will be set equal to the input video pixel aspect ratio of the first input.

h264Settings_flickerAq :: Lens' H264Settings (Maybe H264FlickerAq)

Flicker AQ makes adjustments within each frame to reduce flicker or 'pop' on I-frames. The value to enter in this field depends on the value in the Adaptive quantization field: If you have set the Adaptive quantization field to Auto, MediaLive ignores any value in this field. MediaLive will determine if flicker AQ is appropriate and will apply the appropriate strength. If you have set the Adaptive quantization field to a strength, you can set this field to Enabled or Disabled. Enabled: MediaLive will apply flicker AQ using the specified strength. Disabled: MediaLive won't apply flicker AQ. If you have set the Adaptive quantization to Disabled, MediaLive ignores any value in this field and doesn't apply flicker AQ.

h264Settings_bufSize :: Lens' H264Settings (Maybe Natural)

Size of buffer (HRD buffer model) in bits.

h264Settings_spatialAq :: Lens' H264Settings (Maybe H264SpatialAq)

Spatial AQ makes adjustments within each frame based on spatial variation of content complexity. The value to enter in this field depends on the value in the Adaptive quantization field: If you have set the Adaptive quantization field to Auto, MediaLive ignores any value in this field. MediaLive will determine if spatial AQ is appropriate and will apply the appropriate strength. If you have set the Adaptive quantization field to a strength, you can set this field to Enabled or Disabled. Enabled: MediaLive will apply spatial AQ using the specified strength. Disabled: MediaLive won't apply spatial AQ. If you have set the Adaptive quantization to Disabled, MediaLive ignores any value in this field and doesn't apply spatial AQ.

h264Settings_gopNumBFrames :: Lens' H264Settings (Maybe Natural)

Number of B-frames between reference frames.

h264Settings_fixedAfd :: Lens' H264Settings (Maybe FixedAfd)

Four bit AFD value to write on all frames of video in the output stream. Only valid when afdSignaling is set to 'Fixed'.

h264Settings_softness :: Lens' H264Settings (Maybe Natural)

Softness. Selects quantizer matrix, larger values reduce high-frequency content in the encoded image. If not set to zero, must be greater than 15.

h264Settings_filterSettings :: Lens' H264Settings (Maybe H264FilterSettings)

Optional filters that you can apply to an encode.

h264Settings_bitrate :: Lens' H264Settings (Maybe Natural)

Average bitrate in bits/second. Required when the rate control mode is VBR or CBR. Not used for QVBR. In an MS Smooth output group, each output must have a unique value when its bitrate is rounded down to the nearest multiple of 1000.

h264Settings_framerateDenominator :: Lens' H264Settings (Maybe Natural)

Framerate denominator.

h264Settings_forceFieldPictures :: Lens' H264Settings (Maybe H264ForceFieldPictures)

This setting applies only when scan type is "interlaced." It controls whether coding is performed on a field basis or on a frame basis. (When the video is progressive, the coding is always performed on a frame basis.) enabled: Force MediaLive to code on a field basis, so that odd and even sets of fields are coded separately. disabled: Code the two sets of fields separately (on a field basis) or together (on a frame basis using PAFF), depending on what is most appropriate for the content.

h264Settings_entropyEncoding :: Lens' H264Settings (Maybe H264EntropyEncoding)

Entropy encoding mode. Use cabac (must be in Main or High profile) or cavlc.

h264Settings_framerateControl :: Lens' H264Settings (Maybe H264FramerateControl)

This field indicates how the output video frame rate is specified. If "specified" is selected then the output video frame rate is determined by framerateNumerator and framerateDenominator, else if "initializeFromSource" is selected then the output video frame rate will be set equal to the input video frame rate of the first input.

h264Settings_colorMetadata :: Lens' H264Settings (Maybe H264ColorMetadata)

Includes colorspace metadata in the output.

h264Settings_lookAheadRateControl :: Lens' H264Settings (Maybe H264LookAheadRateControl)

Amount of lookahead. A value of low can decrease latency and memory usage, while high can produce better quality for certain content.

h264Settings_adaptiveQuantization :: Lens' H264Settings (Maybe H264AdaptiveQuantization)

Enables or disables adaptive quantization, which is a technique MediaLive can apply to video on a frame-by-frame basis to produce more compression without losing quality. There are three types of adaptive quantization: flicker, spatial, and temporal. Set the field in one of these ways: Set to Auto. Recommended. For each type of AQ, MediaLive will determine if AQ is needed, and if so, the appropriate strength. Set a strength (a value other than Auto or Disable). This strength will apply to any of the AQ fields that you choose to enable. Set to Disabled to disable all types of adaptive quantization.

h264Settings_framerateNumerator :: Lens' H264Settings (Maybe Natural)

Framerate numerator - framerate is a fraction, e.g. 24000 / 1001 = 23.976 fps.

h264Settings_level :: Lens' H264Settings (Maybe H264Level)

H.264 Level.

h264Settings_gopBReference :: Lens' H264Settings (Maybe H264GopBReference)

Documentation update needed

h264Settings_maxBitrate :: Lens' H264Settings (Maybe Natural)

For QVBR: See the tooltip for Quality level For VBR: Set the maximum bitrate in order to accommodate expected spikes in the complexity of the video.

h264Settings_syntax :: Lens' H264Settings (Maybe H264Syntax)

Produces a bitstream compliant with SMPTE RP-2027.

h264Settings_bufFillPct :: Lens' H264Settings (Maybe Natural)

Percentage of the buffer that should initially be filled (HRD buffer model).

h264Settings_gopClosedCadence :: Lens' H264Settings (Maybe Natural)

Frequency of closed GOPs. In streaming applications, it is recommended that this be set to 1 so a decoder joining mid-stream will receive an IDR frame as quickly as possible. Setting this value to 0 will break output segmenting.

h264Settings_numRefFrames :: Lens' H264Settings (Maybe Natural)

Number of reference frames to use. The encoder may use more than requested if using B-frames and/or interlaced encoding.

h264Settings_parDenominator :: Lens' H264Settings (Maybe Natural)

Pixel Aspect Ratio denominator.