Network.AWS.MediaLive.Types.H265Settings

Description

Documentation

data H265Settings

H265 Settings

See: newH265Settings smart constructor.

Constructors

H265Settings'

Fields

scanType :: Maybe H265ScanType

Sets the scan type of the output to progressive or top-field-first interlaced.

bufSize :: Maybe Natural

Size of buffer (HRD buffer model) in bits.

flickerAq :: Maybe H265FlickerAq

If set to enabled, adjust quantization within each frame to reduce flicker or 'pop' on I-frames.

rateControlMode :: Maybe H265RateControlMode

Rate control mode. QVBR: Quality will match the specified quality level except when it is constrained by the maximum bitrate. Recommended if you or your viewers pay for bandwidth. CBR: Quality varies, depending on the video complexity. Recommended only if you distribute your assets to devices that cannot handle variable bitrates. Multiplex: This rate control mode is only supported (and is required) when the video is being delivered to a MediaLive Multiplex in which case the rate control configuration is controlled by the properties within the Multiplex Program.

alternativeTransferFunction :: Maybe H265AlternativeTransferFunction

Whether or not EML should insert an Alternative Transfer Function SEI message to support backwards compatibility with non-HDR decoders and displays.

slices :: Maybe Natural

Number of slices per picture. Must be less than or equal to the number of macroblock rows for progressive pictures, and less than or equal to half the number of macroblock rows for interlaced pictures. This field is optional; when no value is specified the encoder will choose the number of slices based on encode resolution.

colorMetadata :: Maybe H265ColorMetadata

Includes colorspace metadata in the output.

gopSizeUnits :: Maybe H265GopSizeUnits

Indicates if the gopSize is specified in frames or seconds. If seconds the system will convert the gopSize into a frame count at run time.

fixedAfd :: Maybe FixedAfd

Four bit AFD value to write on all frames of video in the output stream. Only valid when afdSignaling is set to 'Fixed'.

gopSize :: Maybe Double

GOP size (keyframe interval) in units of either frames or seconds per gopSizeUnits. If gopSizeUnits is frames, gopSize must be an integer and must be greater than or equal to 1. If gopSizeUnits is seconds, gopSize must be greater than 0, but need not be an integer.

filterSettings :: Maybe H265FilterSettings

Optional filters that you can apply to an encode.

parNumerator :: Maybe Natural

Pixel Aspect Ratio numerator.

timecodeInsertion :: Maybe H265TimecodeInsertionBehavior

Determines how timecodes should be inserted into the video elementary stream. - 'disabled': Do not include timecodes - 'picTimingSei': Pass through picture timing SEI messages from the source specified in Timecode Config

sceneChangeDetect :: Maybe H265SceneChangeDetect

Scene change detection.

colorSpaceSettings :: Maybe H265ColorSpaceSettings

Color Space settings

minIInterval :: Maybe Natural

Only meaningful if sceneChangeDetect is set to enabled. Defaults to 5 if multiplex rate control is used. Enforces separation between repeated (cadence) I-frames and I-frames inserted by Scene Change Detection. If a scene change I-frame is within I-interval frames of a cadence I-frame, the GOP is shrunk and/or stretched to the scene change I-frame. GOP stretch requires enabling lookahead as well as setting I-interval. The normal cadence resumes for the next GOP. Note: Maximum GOP stretch = GOP size + Min-I-interval - 1

qvbrQualityLevel :: Maybe Natural

Controls the target quality for the video encode. Applies only when the rate control mode is QVBR. Set values for the QVBR quality level field and Max bitrate field that suit your most important viewing devices. Recommended values are: - Primary screen: Quality level: 8 to 10. Max bitrate: 4M - PC or tablet: Quality level: 7. Max bitrate: 1.5M to 3M - Smartphone: Quality level: 6. Max bitrate: 1M to 1.5M

gopClosedCadence :: Maybe Natural

Frequency of closed GOPs. In streaming applications, it is recommended that this be set to 1 so a decoder joining mid-stream will receive an IDR frame as quickly as possible. Setting this value to 0 will break output segmenting.

parDenominator :: Maybe Natural

Pixel Aspect Ratio denominator.

maxBitrate :: Maybe Natural

For QVBR: See the tooltip for Quality level

level :: Maybe H265Level

H.265 Level.

lookAheadRateControl :: Maybe H265LookAheadRateControl

Amount of lookahead. A value of low can decrease latency and memory usage, while high can produce better quality for certain content.

profile :: Maybe H265Profile

H.265 Profile.

adaptiveQuantization :: Maybe H265AdaptiveQuantization

Adaptive quantization. Allows intra-frame quantizers to vary to improve visual quality.

bitrate :: Maybe Natural

Average bitrate in bits/second. Required when the rate control mode is VBR or CBR. Not used for QVBR. In an MS Smooth output group, each output must have a unique value when its bitrate is rounded down to the nearest multiple of 1000.

afdSignaling :: Maybe AfdSignaling

Indicates that AFD values will be written into the output stream. If afdSignaling is "auto", the system will try to preserve the input AFD value (in cases where multiple AFD values are valid). If set to "fixed", the AFD value will be the value configured in the fixedAfd parameter.

tier :: Maybe H265Tier

H.265 Tier.

framerateNumerator :: Natural

Framerate numerator - framerate is a fraction, e.g. 24000 / 1001 = 23.976 fps.

framerateDenominator :: Natural

Framerate denominator.

newH265Settings

Arguments

:: Natural
$sel:framerateNumerator:H265Settings'

-> Natural
$sel:framerateDenominator:H265Settings'

-> H265Settings

Create a value of H265Settings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scanType:H265Settings', h265Settings_scanType - Sets the scan type of the output to progressive or top-field-first interlaced.

$sel:bufSize:H265Settings', h265Settings_bufSize - Size of buffer (HRD buffer model) in bits.

$sel:flickerAq:H265Settings', h265Settings_flickerAq - If set to enabled, adjust quantization within each frame to reduce flicker or 'pop' on I-frames.

$sel:rateControlMode:H265Settings', h265Settings_rateControlMode - Rate control mode. QVBR: Quality will match the specified quality level except when it is constrained by the maximum bitrate. Recommended if you or your viewers pay for bandwidth. CBR: Quality varies, depending on the video complexity. Recommended only if you distribute your assets to devices that cannot handle variable bitrates. Multiplex: This rate control mode is only supported (and is required) when the video is being delivered to a MediaLive Multiplex in which case the rate control configuration is controlled by the properties within the Multiplex Program.

$sel:alternativeTransferFunction:H265Settings', h265Settings_alternativeTransferFunction - Whether or not EML should insert an Alternative Transfer Function SEI message to support backwards compatibility with non-HDR decoders and displays.

$sel:slices:H265Settings', h265Settings_slices - Number of slices per picture. Must be less than or equal to the number of macroblock rows for progressive pictures, and less than or equal to half the number of macroblock rows for interlaced pictures. This field is optional; when no value is specified the encoder will choose the number of slices based on encode resolution.

$sel:colorMetadata:H265Settings', h265Settings_colorMetadata - Includes colorspace metadata in the output.

$sel:gopSizeUnits:H265Settings', h265Settings_gopSizeUnits - Indicates if the gopSize is specified in frames or seconds. If seconds the system will convert the gopSize into a frame count at run time.

$sel:fixedAfd:H265Settings', h265Settings_fixedAfd - Four bit AFD value to write on all frames of video in the output stream. Only valid when afdSignaling is set to 'Fixed'.

$sel:gopSize:H265Settings', h265Settings_gopSize - GOP size (keyframe interval) in units of either frames or seconds per gopSizeUnits. If gopSizeUnits is frames, gopSize must be an integer and must be greater than or equal to 1. If gopSizeUnits is seconds, gopSize must be greater than 0, but need not be an integer.

$sel:filterSettings:H265Settings', h265Settings_filterSettings - Optional filters that you can apply to an encode.

$sel:parNumerator:H265Settings', h265Settings_parNumerator - Pixel Aspect Ratio numerator.

$sel:timecodeInsertion:H265Settings', h265Settings_timecodeInsertion - Determines how timecodes should be inserted into the video elementary stream. - 'disabled': Do not include timecodes - 'picTimingSei': Pass through picture timing SEI messages from the source specified in Timecode Config

$sel:sceneChangeDetect:H265Settings', h265Settings_sceneChangeDetect - Scene change detection.

$sel:colorSpaceSettings:H265Settings', h265Settings_colorSpaceSettings - Color Space settings

$sel:minIInterval:H265Settings', h265Settings_minIInterval - Only meaningful if sceneChangeDetect is set to enabled. Defaults to 5 if multiplex rate control is used. Enforces separation between repeated (cadence) I-frames and I-frames inserted by Scene Change Detection. If a scene change I-frame is within I-interval frames of a cadence I-frame, the GOP is shrunk and/or stretched to the scene change I-frame. GOP stretch requires enabling lookahead as well as setting I-interval. The normal cadence resumes for the next GOP. Note: Maximum GOP stretch = GOP size + Min-I-interval - 1

$sel:qvbrQualityLevel:H265Settings', h265Settings_qvbrQualityLevel - Controls the target quality for the video encode. Applies only when the rate control mode is QVBR. Set values for the QVBR quality level field and Max bitrate field that suit your most important viewing devices. Recommended values are: - Primary screen: Quality level: 8 to 10. Max bitrate: 4M - PC or tablet: Quality level: 7. Max bitrate: 1.5M to 3M - Smartphone: Quality level: 6. Max bitrate: 1M to 1.5M

$sel:gopClosedCadence:H265Settings', h265Settings_gopClosedCadence - Frequency of closed GOPs. In streaming applications, it is recommended that this be set to 1 so a decoder joining mid-stream will receive an IDR frame as quickly as possible. Setting this value to 0 will break output segmenting.

$sel:parDenominator:H265Settings', h265Settings_parDenominator - Pixel Aspect Ratio denominator.

$sel:maxBitrate:H265Settings', h265Settings_maxBitrate - For QVBR: See the tooltip for Quality level

$sel:level:H265Settings', h265Settings_level - H.265 Level.

$sel:lookAheadRateControl:H265Settings', h265Settings_lookAheadRateControl - Amount of lookahead. A value of low can decrease latency and memory usage, while high can produce better quality for certain content.

$sel:profile:H265Settings', h265Settings_profile - H.265 Profile.

$sel:adaptiveQuantization:H265Settings', h265Settings_adaptiveQuantization - Adaptive quantization. Allows intra-frame quantizers to vary to improve visual quality.

$sel:bitrate:H265Settings', h265Settings_bitrate - Average bitrate in bits/second. Required when the rate control mode is VBR or CBR. Not used for QVBR. In an MS Smooth output group, each output must have a unique value when its bitrate is rounded down to the nearest multiple of 1000.

$sel:afdSignaling:H265Settings', h265Settings_afdSignaling - Indicates that AFD values will be written into the output stream. If afdSignaling is "auto", the system will try to preserve the input AFD value (in cases where multiple AFD values are valid). If set to "fixed", the AFD value will be the value configured in the fixedAfd parameter.

$sel:tier:H265Settings', h265Settings_tier - H.265 Tier.

$sel:framerateNumerator:H265Settings', h265Settings_framerateNumerator - Framerate numerator - framerate is a fraction, e.g. 24000 / 1001 = 23.976 fps.

$sel:framerateDenominator:H265Settings', h265Settings_framerateDenominator - Framerate denominator.

h265Settings_scanType :: Lens' H265Settings (Maybe H265ScanType)

Sets the scan type of the output to progressive or top-field-first interlaced.

h265Settings_bufSize :: Lens' H265Settings (Maybe Natural)

Size of buffer (HRD buffer model) in bits.

h265Settings_flickerAq :: Lens' H265Settings (Maybe H265FlickerAq)

If set to enabled, adjust quantization within each frame to reduce flicker or 'pop' on I-frames.

h265Settings_rateControlMode :: Lens' H265Settings (Maybe H265RateControlMode)

Rate control mode. QVBR: Quality will match the specified quality level except when it is constrained by the maximum bitrate. Recommended if you or your viewers pay for bandwidth. CBR: Quality varies, depending on the video complexity. Recommended only if you distribute your assets to devices that cannot handle variable bitrates. Multiplex: This rate control mode is only supported (and is required) when the video is being delivered to a MediaLive Multiplex in which case the rate control configuration is controlled by the properties within the Multiplex Program.

h265Settings_alternativeTransferFunction :: Lens' H265Settings (Maybe H265AlternativeTransferFunction)

Whether or not EML should insert an Alternative Transfer Function SEI message to support backwards compatibility with non-HDR decoders and displays.

h265Settings_slices :: Lens' H265Settings (Maybe Natural)

Number of slices per picture. Must be less than or equal to the number of macroblock rows for progressive pictures, and less than or equal to half the number of macroblock rows for interlaced pictures. This field is optional; when no value is specified the encoder will choose the number of slices based on encode resolution.

h265Settings_colorMetadata :: Lens' H265Settings (Maybe H265ColorMetadata)

Includes colorspace metadata in the output.

h265Settings_gopSizeUnits :: Lens' H265Settings (Maybe H265GopSizeUnits)

Indicates if the gopSize is specified in frames or seconds. If seconds the system will convert the gopSize into a frame count at run time.

h265Settings_fixedAfd :: Lens' H265Settings (Maybe FixedAfd)

Four bit AFD value to write on all frames of video in the output stream. Only valid when afdSignaling is set to 'Fixed'.

h265Settings_gopSize :: Lens' H265Settings (Maybe Double)

GOP size (keyframe interval) in units of either frames or seconds per gopSizeUnits. If gopSizeUnits is frames, gopSize must be an integer and must be greater than or equal to 1. If gopSizeUnits is seconds, gopSize must be greater than 0, but need not be an integer.

h265Settings_filterSettings :: Lens' H265Settings (Maybe H265FilterSettings)

Optional filters that you can apply to an encode.

h265Settings_parNumerator :: Lens' H265Settings (Maybe Natural)

Pixel Aspect Ratio numerator.

h265Settings_timecodeInsertion :: Lens' H265Settings (Maybe H265TimecodeInsertionBehavior)

Determines how timecodes should be inserted into the video elementary stream. - 'disabled': Do not include timecodes - 'picTimingSei': Pass through picture timing SEI messages from the source specified in Timecode Config

h265Settings_sceneChangeDetect :: Lens' H265Settings (Maybe H265SceneChangeDetect)

Scene change detection.

h265Settings_colorSpaceSettings :: Lens' H265Settings (Maybe H265ColorSpaceSettings)

Color Space settings

h265Settings_minIInterval :: Lens' H265Settings (Maybe Natural)

Only meaningful if sceneChangeDetect is set to enabled. Defaults to 5 if multiplex rate control is used. Enforces separation between repeated (cadence) I-frames and I-frames inserted by Scene Change Detection. If a scene change I-frame is within I-interval frames of a cadence I-frame, the GOP is shrunk and/or stretched to the scene change I-frame. GOP stretch requires enabling lookahead as well as setting I-interval. The normal cadence resumes for the next GOP. Note: Maximum GOP stretch = GOP size + Min-I-interval - 1

h265Settings_qvbrQualityLevel :: Lens' H265Settings (Maybe Natural)

Controls the target quality for the video encode. Applies only when the rate control mode is QVBR. Set values for the QVBR quality level field and Max bitrate field that suit your most important viewing devices. Recommended values are: - Primary screen: Quality level: 8 to 10. Max bitrate: 4M - PC or tablet: Quality level: 7. Max bitrate: 1.5M to 3M - Smartphone: Quality level: 6. Max bitrate: 1M to 1.5M

h265Settings_gopClosedCadence :: Lens' H265Settings (Maybe Natural)

Frequency of closed GOPs. In streaming applications, it is recommended that this be set to 1 so a decoder joining mid-stream will receive an IDR frame as quickly as possible. Setting this value to 0 will break output segmenting.

h265Settings_parDenominator :: Lens' H265Settings (Maybe Natural)

Pixel Aspect Ratio denominator.

h265Settings_maxBitrate :: Lens' H265Settings (Maybe Natural)

For QVBR: See the tooltip for Quality level

h265Settings_level :: Lens' H265Settings (Maybe H265Level)

H.265 Level.

h265Settings_lookAheadRateControl :: Lens' H265Settings (Maybe H265LookAheadRateControl)

Amount of lookahead. A value of low can decrease latency and memory usage, while high can produce better quality for certain content.

h265Settings_profile :: Lens' H265Settings (Maybe H265Profile)

H.265 Profile.

h265Settings_adaptiveQuantization :: Lens' H265Settings (Maybe H265AdaptiveQuantization)

Adaptive quantization. Allows intra-frame quantizers to vary to improve visual quality.

h265Settings_bitrate :: Lens' H265Settings (Maybe Natural)

Average bitrate in bits/second. Required when the rate control mode is VBR or CBR. Not used for QVBR. In an MS Smooth output group, each output must have a unique value when its bitrate is rounded down to the nearest multiple of 1000.

h265Settings_afdSignaling :: Lens' H265Settings (Maybe AfdSignaling)

Indicates that AFD values will be written into the output stream. If afdSignaling is "auto", the system will try to preserve the input AFD value (in cases where multiple AFD values are valid). If set to "fixed", the AFD value will be the value configured in the fixedAfd parameter.

h265Settings_tier :: Lens' H265Settings (Maybe H265Tier)

H.265 Tier.

h265Settings_framerateNumerator :: Lens' H265Settings Natural

Framerate numerator - framerate is a fraction, e.g. 24000 / 1001 = 23.976 fps.

h265Settings_framerateDenominator :: Lens' H265Settings Natural

Framerate denominator.