Network.AWS.MediaLive.Types.HlsGroupSettings

Description

Documentation

data HlsGroupSettings

Hls Group Settings

See: newHlsGroupSettings smart constructor.

Constructors

HlsGroupSettings'

Fields

segmentLength :: Maybe Natural

Length of MPEG-2 Transport Stream segments to create (in seconds). Note that segments will end on the next keyframe after this number of seconds, so actual segment length may be longer.

timedMetadataId3Frame :: Maybe HlsTimedMetadataId3Frame

Indicates ID3 frame that has the timecode.

ivInManifest :: Maybe HlsIvInManifest

For use with encryptionType. The IV (Initialization Vector) is a 128-bit number used in conjunction with the key for encrypting blocks. If set to "include", IV is listed in the manifest, otherwise the IV is not in the manifest.

outputSelection :: Maybe HlsOutputSelection

MANIFESTS_AND_SEGMENTS: Generates manifests (master manifest, if applicable, and media manifests) for this output group. VARIANT_MANIFESTS_AND_SEGMENTS: Generates media manifests for this output group, but not a master manifest. SEGMENTS_ONLY: Does not generate any manifests for this output group.

timedMetadataId3Period :: Maybe Natural

Timed Metadata interval in seconds.

encryptionType :: Maybe HlsEncryptionType

Encrypts the segments with the given encryption scheme. Exclude this parameter if no encryption is desired.

directoryStructure :: Maybe HlsDirectoryStructure

Place segments in subdirectories.

keyFormat :: Maybe Text

The value specifies how the key is represented in the resource identified by the URI. If parameter is absent, an implicit value of "identity" is used. A reverse DNS string can also be given.

adMarkers :: Maybe [HlsAdMarkers]

Choose one or more ad marker types to pass SCTE35 signals through to this group of Apple HLS outputs.

manifestCompression :: Maybe HlsManifestCompression

When set to gzip, compresses HLS playlist.

constantIv :: Maybe Text

For use with encryptionType. This is a 128-bit, 16-byte hex value represented by a 32-character text string. If ivSource is set to "explicit" then this parameter is required and is used as the IV for encryption.

codecSpecification :: Maybe HlsCodecSpecification

Specification to use (RFC-6381 or the default RFC-4281) during m3u8 playlist generation.

hlsCdnSettings :: Maybe HlsCdnSettings

Parameters that control interactions with the CDN.

mode :: Maybe HlsMode

If "vod", all segments are indexed and kept permanently in the destination and manifest. If "live", only the number segments specified in keepSegments and indexNSegments are kept; newer segments replace older segments, which may prevent players from rewinding all the way to the beginning of the event. VOD mode uses HLS EXT-X-PLAYLIST-TYPE of EVENT while the channel is running, converting it to a "VOD" type manifest on completion of the stream.

streamInfResolution :: Maybe HlsStreamInfResolution

Include or exclude RESOLUTION attribute for video in EXT-X-STREAM-INF tag of variant manifest.

indexNSegments :: Maybe Natural

Applies only if Mode field is LIVE. Specifies the maximum number of segments in the media manifest file. After this maximum, older segments are removed from the media manifest. This number must be smaller than the number in the Keep Segments field.

redundantManifest :: Maybe HlsRedundantManifest

ENABLED: The master manifest (.m3u8 file) for each pipeline includes information about both pipelines: first its own media files, then the media files of the other pipeline. This feature allows playout device that support stale manifest detection to switch from one manifest to the other, when the current manifest seems to be stale. There are still two destinations and two master manifests, but both master manifests reference the media files from both pipelines. DISABLED: The master manifest (.m3u8 file) for each pipeline includes information about its own pipeline only. For an HLS output group with MediaPackage as the destination, the DISABLED behavior is always followed. MediaPackage regenerates the manifests it serves to players so a redundant manifest from MediaLive is irrelevant.

segmentationMode :: Maybe HlsSegmentationMode

useInputSegmentation has been deprecated. The configured segment size is always used.

programDateTime :: Maybe HlsProgramDateTime

Includes or excludes EXT-X-PROGRAM-DATE-TIME tag in .m3u8 manifest files. The value is calculated as follows: either the program date and time are initialized using the input timecode source, or the time is initialized using the input timecode source and the date is initialized using the timestampOffset.

iFrameOnlyPlaylists :: Maybe IFrameOnlyPlaylistType

DISABLED: Do not create an I-frame-only manifest, but do create the master and media manifests (according to the Output Selection field). STANDARD: Create an I-frame-only manifest for each output that contains video, as well as the other manifests (according to the Output Selection field). The I-frame manifest contains a #EXT-X-I-FRAMES-ONLY tag to indicate it is I-frame only, and one or more #EXT-X-BYTERANGE entries identifying the I-frame position. For example, #EXT-X-BYTERANGE:160364@1461888"

tsFileMode :: Maybe HlsTsFileMode

SEGMENTED_FILES: Emit the program as segments - multiple .ts media files. SINGLE_FILE: Applies only if Mode field is VOD. Emit the program as a single .ts media file. The media manifest includes #EXT-X-BYTERANGE tags to index segments for playback. A typical use for this value is when sending the output to AWS Elemental MediaConvert, which can accept only a single media file. Playback while the channel is running is not guaranteed due to HTTP server caching.

segmentsPerSubdirectory :: Maybe Natural

Number of segments to write to a subdirectory before starting a new one. directoryStructure must be subdirectoryPerStream for this setting to have an effect.

discontinuityTags :: Maybe HlsDiscontinuityTags

Specifies whether to insert EXT-X-DISCONTINUITY tags in the HLS child manifests for this output group. Typically, choose Insert because these tags are required in the manifest (according to the HLS specification) and serve an important purpose. Choose Never Insert only if the downstream system is doing real-time failover (without using the MediaLive automatic failover feature) and only if that downstream system has advised you to exclude the tags.

baseUrlContent :: Maybe Text

A partial URI prefix that will be prepended to each output in the media .m3u8 file. Can be used if base manifest is delivered from a different URL than the main .m3u8 file.

hlsId3SegmentTagging :: Maybe HlsId3SegmentTaggingState

State of HLS ID3 Segment Tagging

baseUrlManifest :: Maybe Text

A partial URI prefix that will be prepended to each output in the media .m3u8 file. Can be used if base manifest is delivered from a different URL than the main .m3u8 file.

incompleteSegmentBehavior :: Maybe HlsIncompleteSegmentBehavior

Specifies whether to include the final (incomplete) segment in the media output when the pipeline stops producing output because of a channel stop, a channel pause or a loss of input to the pipeline. Auto means that MediaLive decides whether to include the final segment, depending on the channel class and the types of output groups. Suppress means to never include the incomplete segment. We recommend you choose Auto and let MediaLive control the behavior.

baseUrlContent1 :: Maybe Text

Optional. One value per output group. This field is required only if you are completing Base URL content A, and the downstream system has notified you that the media files for pipeline 1 of all outputs are in a location different from the media files for pipeline 0.

inputLossAction :: Maybe InputLossActionForHlsOut

Parameter that control output group behavior on input loss.

keyProviderSettings :: Maybe KeyProviderSettings

The key provider settings.

keepSegments :: Maybe Natural

Applies only if Mode field is LIVE. Specifies the number of media segments to retain in the destination directory. This number should be bigger than indexNSegments (Num segments). We recommend (value = (2 x indexNsegments) + 1). If this "keep segments" number is too low, the following might happen: the player is still reading a media manifest file that lists this segment, but that segment has been removed from the destination directory (as directed by indexNSegments). This situation would result in a 404 HTTP error on the player.

captionLanguageMappings :: Maybe [CaptionLanguageMapping]

Mapping of up to 4 caption channels to caption languages. Is only meaningful if captionLanguageSetting is set to "insert".

timestampDeltaMilliseconds :: Maybe Natural

Provides an extra millisecond delta offset to fine tune the timestamps.

baseUrlManifest1 :: Maybe Text

Optional. One value per output group. Complete this field only if you are completing Base URL manifest A, and the downstream system has notified you that the child manifest files for pipeline 1 of all outputs are in a location different from the child manifest files for pipeline 0.

programDateTimePeriod :: Maybe Natural

Period of insertion of EXT-X-PROGRAM-DATE-TIME entry, in seconds.

clientCache :: Maybe HlsClientCache

When set to "disabled", sets the #EXT-X-ALLOW-CACHE:no tag in the manifest, which prevents clients from saving media segments for later replay.

keyFormatVersions :: Maybe Text

Either a single positive integer version value or a slash delimited list of version values (1/2/3).

minSegmentLength :: Maybe Natural

When set, minimumSegmentLength is enforced by looking ahead and back within the specified range for a nearby avail and extending the segment size if needed.

ivSource :: Maybe HlsIvSource

For use with encryptionType. The IV (Initialization Vector) is a 128-bit number used in conjunction with the key for encrypting blocks. If this setting is "followsSegmentNumber", it will cause the IV to change every segment (to match the segment number). If this is set to "explicit", you must enter a constantIv value.

captionLanguageSetting :: Maybe HlsCaptionLanguageSetting

Applies only to 608 Embedded output captions. insert: Include CLOSED-CAPTIONS lines in the manifest. Specify at least one language in the CC1 Language Code field. One CLOSED-CAPTION line is added for each Language Code you specify. Make sure to specify the languages in the order in which they appear in the original source (if the source is embedded format) or the order of the caption selectors (if the source is other than embedded). Otherwise, languages in the manifest will not match up properly with the output captions. none: Include CLOSED-CAPTIONS=NONE line in the manifest. omit: Omit any CLOSED-CAPTIONS line from the manifest.

manifestDurationFormat :: Maybe HlsManifestDurationFormat

Indicates whether the output manifest should use floating point or integer values for segment duration.

destination :: OutputLocationRef

A directory or HTTP destination for the HLS segments, manifest files, and encryption keys (if enabled).

newHlsGroupSettings

Arguments

:: OutputLocationRef
$sel:destination:HlsGroupSettings'

-> HlsGroupSettings

Create a value of HlsGroupSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:segmentLength:HlsGroupSettings', hlsGroupSettings_segmentLength - Length of MPEG-2 Transport Stream segments to create (in seconds). Note that segments will end on the next keyframe after this number of seconds, so actual segment length may be longer.

$sel:timedMetadataId3Frame:HlsGroupSettings', hlsGroupSettings_timedMetadataId3Frame - Indicates ID3 frame that has the timecode.

$sel:ivInManifest:HlsGroupSettings', hlsGroupSettings_ivInManifest - For use with encryptionType. The IV (Initialization Vector) is a 128-bit number used in conjunction with the key for encrypting blocks. If set to "include", IV is listed in the manifest, otherwise the IV is not in the manifest.

$sel:outputSelection:HlsGroupSettings', hlsGroupSettings_outputSelection - MANIFESTS_AND_SEGMENTS: Generates manifests (master manifest, if applicable, and media manifests) for this output group. VARIANT_MANIFESTS_AND_SEGMENTS: Generates media manifests for this output group, but not a master manifest. SEGMENTS_ONLY: Does not generate any manifests for this output group.

$sel:timedMetadataId3Period:HlsGroupSettings', hlsGroupSettings_timedMetadataId3Period - Timed Metadata interval in seconds.

$sel:encryptionType:HlsGroupSettings', hlsGroupSettings_encryptionType - Encrypts the segments with the given encryption scheme. Exclude this parameter if no encryption is desired.

$sel:directoryStructure:HlsGroupSettings', hlsGroupSettings_directoryStructure - Place segments in subdirectories.

$sel:keyFormat:HlsGroupSettings', hlsGroupSettings_keyFormat - The value specifies how the key is represented in the resource identified by the URI. If parameter is absent, an implicit value of "identity" is used. A reverse DNS string can also be given.

$sel:adMarkers:HlsGroupSettings', hlsGroupSettings_adMarkers - Choose one or more ad marker types to pass SCTE35 signals through to this group of Apple HLS outputs.

$sel:manifestCompression:HlsGroupSettings', hlsGroupSettings_manifestCompression - When set to gzip, compresses HLS playlist.

$sel:constantIv:HlsGroupSettings', hlsGroupSettings_constantIv - For use with encryptionType. This is a 128-bit, 16-byte hex value represented by a 32-character text string. If ivSource is set to "explicit" then this parameter is required and is used as the IV for encryption.

$sel:codecSpecification:HlsGroupSettings', hlsGroupSettings_codecSpecification - Specification to use (RFC-6381 or the default RFC-4281) during m3u8 playlist generation.

$sel:hlsCdnSettings:HlsGroupSettings', hlsGroupSettings_hlsCdnSettings - Parameters that control interactions with the CDN.

$sel:mode:HlsGroupSettings', hlsGroupSettings_mode - If "vod", all segments are indexed and kept permanently in the destination and manifest. If "live", only the number segments specified in keepSegments and indexNSegments are kept; newer segments replace older segments, which may prevent players from rewinding all the way to the beginning of the event. VOD mode uses HLS EXT-X-PLAYLIST-TYPE of EVENT while the channel is running, converting it to a "VOD" type manifest on completion of the stream.

$sel:streamInfResolution:HlsGroupSettings', hlsGroupSettings_streamInfResolution - Include or exclude RESOLUTION attribute for video in EXT-X-STREAM-INF tag of variant manifest.

$sel:indexNSegments:HlsGroupSettings', hlsGroupSettings_indexNSegments - Applies only if Mode field is LIVE. Specifies the maximum number of segments in the media manifest file. After this maximum, older segments are removed from the media manifest. This number must be smaller than the number in the Keep Segments field.

$sel:redundantManifest:HlsGroupSettings', hlsGroupSettings_redundantManifest - ENABLED: The master manifest (.m3u8 file) for each pipeline includes information about both pipelines: first its own media files, then the media files of the other pipeline. This feature allows playout device that support stale manifest detection to switch from one manifest to the other, when the current manifest seems to be stale. There are still two destinations and two master manifests, but both master manifests reference the media files from both pipelines. DISABLED: The master manifest (.m3u8 file) for each pipeline includes information about its own pipeline only. For an HLS output group with MediaPackage as the destination, the DISABLED behavior is always followed. MediaPackage regenerates the manifests it serves to players so a redundant manifest from MediaLive is irrelevant.

$sel:segmentationMode:HlsGroupSettings', hlsGroupSettings_segmentationMode - useInputSegmentation has been deprecated. The configured segment size is always used.

$sel:programDateTime:HlsGroupSettings', hlsGroupSettings_programDateTime - Includes or excludes EXT-X-PROGRAM-DATE-TIME tag in .m3u8 manifest files. The value is calculated as follows: either the program date and time are initialized using the input timecode source, or the time is initialized using the input timecode source and the date is initialized using the timestampOffset.

$sel:iFrameOnlyPlaylists:HlsGroupSettings', hlsGroupSettings_iFrameOnlyPlaylists - DISABLED: Do not create an I-frame-only manifest, but do create the master and media manifests (according to the Output Selection field). STANDARD: Create an I-frame-only manifest for each output that contains video, as well as the other manifests (according to the Output Selection field). The I-frame manifest contains a #EXT-X-I-FRAMES-ONLY tag to indicate it is I-frame only, and one or more #EXT-X-BYTERANGE entries identifying the I-frame position. For example, #EXT-X-BYTERANGE:160364@1461888"

$sel:tsFileMode:HlsGroupSettings', hlsGroupSettings_tsFileMode - SEGMENTED_FILES: Emit the program as segments - multiple .ts media files. SINGLE_FILE: Applies only if Mode field is VOD. Emit the program as a single .ts media file. The media manifest includes #EXT-X-BYTERANGE tags to index segments for playback. A typical use for this value is when sending the output to AWS Elemental MediaConvert, which can accept only a single media file. Playback while the channel is running is not guaranteed due to HTTP server caching.

$sel:segmentsPerSubdirectory:HlsGroupSettings', hlsGroupSettings_segmentsPerSubdirectory - Number of segments to write to a subdirectory before starting a new one. directoryStructure must be subdirectoryPerStream for this setting to have an effect.

$sel:discontinuityTags:HlsGroupSettings', hlsGroupSettings_discontinuityTags - Specifies whether to insert EXT-X-DISCONTINUITY tags in the HLS child manifests for this output group. Typically, choose Insert because these tags are required in the manifest (according to the HLS specification) and serve an important purpose. Choose Never Insert only if the downstream system is doing real-time failover (without using the MediaLive automatic failover feature) and only if that downstream system has advised you to exclude the tags.

$sel:baseUrlContent:HlsGroupSettings', hlsGroupSettings_baseUrlContent - A partial URI prefix that will be prepended to each output in the media .m3u8 file. Can be used if base manifest is delivered from a different URL than the main .m3u8 file.

$sel:hlsId3SegmentTagging:HlsGroupSettings', hlsGroupSettings_hlsId3SegmentTagging - State of HLS ID3 Segment Tagging

$sel:baseUrlManifest:HlsGroupSettings', hlsGroupSettings_baseUrlManifest - A partial URI prefix that will be prepended to each output in the media .m3u8 file. Can be used if base manifest is delivered from a different URL than the main .m3u8 file.

$sel:incompleteSegmentBehavior:HlsGroupSettings', hlsGroupSettings_incompleteSegmentBehavior - Specifies whether to include the final (incomplete) segment in the media output when the pipeline stops producing output because of a channel stop, a channel pause or a loss of input to the pipeline. Auto means that MediaLive decides whether to include the final segment, depending on the channel class and the types of output groups. Suppress means to never include the incomplete segment. We recommend you choose Auto and let MediaLive control the behavior.

$sel:baseUrlContent1:HlsGroupSettings', hlsGroupSettings_baseUrlContent1 - Optional. One value per output group. This field is required only if you are completing Base URL content A, and the downstream system has notified you that the media files for pipeline 1 of all outputs are in a location different from the media files for pipeline 0.

$sel:inputLossAction:HlsGroupSettings', hlsGroupSettings_inputLossAction - Parameter that control output group behavior on input loss.

$sel:keyProviderSettings:HlsGroupSettings', hlsGroupSettings_keyProviderSettings - The key provider settings.

$sel:keepSegments:HlsGroupSettings', hlsGroupSettings_keepSegments - Applies only if Mode field is LIVE. Specifies the number of media segments to retain in the destination directory. This number should be bigger than indexNSegments (Num segments). We recommend (value = (2 x indexNsegments) + 1). If this "keep segments" number is too low, the following might happen: the player is still reading a media manifest file that lists this segment, but that segment has been removed from the destination directory (as directed by indexNSegments). This situation would result in a 404 HTTP error on the player.

$sel:captionLanguageMappings:HlsGroupSettings', hlsGroupSettings_captionLanguageMappings - Mapping of up to 4 caption channels to caption languages. Is only meaningful if captionLanguageSetting is set to "insert".

$sel:timestampDeltaMilliseconds:HlsGroupSettings', hlsGroupSettings_timestampDeltaMilliseconds - Provides an extra millisecond delta offset to fine tune the timestamps.

$sel:baseUrlManifest1:HlsGroupSettings', hlsGroupSettings_baseUrlManifest1 - Optional. One value per output group. Complete this field only if you are completing Base URL manifest A, and the downstream system has notified you that the child manifest files for pipeline 1 of all outputs are in a location different from the child manifest files for pipeline 0.

$sel:programDateTimePeriod:HlsGroupSettings', hlsGroupSettings_programDateTimePeriod - Period of insertion of EXT-X-PROGRAM-DATE-TIME entry, in seconds.

$sel:clientCache:HlsGroupSettings', hlsGroupSettings_clientCache - When set to "disabled", sets the #EXT-X-ALLOW-CACHE:no tag in the manifest, which prevents clients from saving media segments for later replay.

$sel:keyFormatVersions:HlsGroupSettings', hlsGroupSettings_keyFormatVersions - Either a single positive integer version value or a slash delimited list of version values (1/2/3).

$sel:minSegmentLength:HlsGroupSettings', hlsGroupSettings_minSegmentLength - When set, minimumSegmentLength is enforced by looking ahead and back within the specified range for a nearby avail and extending the segment size if needed.

$sel:ivSource:HlsGroupSettings', hlsGroupSettings_ivSource - For use with encryptionType. The IV (Initialization Vector) is a 128-bit number used in conjunction with the key for encrypting blocks. If this setting is "followsSegmentNumber", it will cause the IV to change every segment (to match the segment number). If this is set to "explicit", you must enter a constantIv value.

$sel:captionLanguageSetting:HlsGroupSettings', hlsGroupSettings_captionLanguageSetting - Applies only to 608 Embedded output captions. insert: Include CLOSED-CAPTIONS lines in the manifest. Specify at least one language in the CC1 Language Code field. One CLOSED-CAPTION line is added for each Language Code you specify. Make sure to specify the languages in the order in which they appear in the original source (if the source is embedded format) or the order of the caption selectors (if the source is other than embedded). Otherwise, languages in the manifest will not match up properly with the output captions. none: Include CLOSED-CAPTIONS=NONE line in the manifest. omit: Omit any CLOSED-CAPTIONS line from the manifest.

$sel:manifestDurationFormat:HlsGroupSettings', hlsGroupSettings_manifestDurationFormat - Indicates whether the output manifest should use floating point or integer values for segment duration.

$sel:destination:HlsGroupSettings', hlsGroupSettings_destination - A directory or HTTP destination for the HLS segments, manifest files, and encryption keys (if enabled).

hlsGroupSettings_segmentLength :: Lens' HlsGroupSettings (Maybe Natural)

Length of MPEG-2 Transport Stream segments to create (in seconds). Note that segments will end on the next keyframe after this number of seconds, so actual segment length may be longer.

hlsGroupSettings_timedMetadataId3Frame :: Lens' HlsGroupSettings (Maybe HlsTimedMetadataId3Frame)

Indicates ID3 frame that has the timecode.

hlsGroupSettings_ivInManifest :: Lens' HlsGroupSettings (Maybe HlsIvInManifest)

For use with encryptionType. The IV (Initialization Vector) is a 128-bit number used in conjunction with the key for encrypting blocks. If set to "include", IV is listed in the manifest, otherwise the IV is not in the manifest.

hlsGroupSettings_outputSelection :: Lens' HlsGroupSettings (Maybe HlsOutputSelection)

MANIFESTS_AND_SEGMENTS: Generates manifests (master manifest, if applicable, and media manifests) for this output group. VARIANT_MANIFESTS_AND_SEGMENTS: Generates media manifests for this output group, but not a master manifest. SEGMENTS_ONLY: Does not generate any manifests for this output group.

hlsGroupSettings_timedMetadataId3Period :: Lens' HlsGroupSettings (Maybe Natural)

Timed Metadata interval in seconds.

hlsGroupSettings_encryptionType :: Lens' HlsGroupSettings (Maybe HlsEncryptionType)

Encrypts the segments with the given encryption scheme. Exclude this parameter if no encryption is desired.

hlsGroupSettings_directoryStructure :: Lens' HlsGroupSettings (Maybe HlsDirectoryStructure)

Place segments in subdirectories.

hlsGroupSettings_keyFormat :: Lens' HlsGroupSettings (Maybe Text)

The value specifies how the key is represented in the resource identified by the URI. If parameter is absent, an implicit value of "identity" is used. A reverse DNS string can also be given.

hlsGroupSettings_adMarkers :: Lens' HlsGroupSettings (Maybe [HlsAdMarkers])

Choose one or more ad marker types to pass SCTE35 signals through to this group of Apple HLS outputs.

hlsGroupSettings_manifestCompression :: Lens' HlsGroupSettings (Maybe HlsManifestCompression)

When set to gzip, compresses HLS playlist.

hlsGroupSettings_constantIv :: Lens' HlsGroupSettings (Maybe Text)

For use with encryptionType. This is a 128-bit, 16-byte hex value represented by a 32-character text string. If ivSource is set to "explicit" then this parameter is required and is used as the IV for encryption.

hlsGroupSettings_codecSpecification :: Lens' HlsGroupSettings (Maybe HlsCodecSpecification)

Specification to use (RFC-6381 or the default RFC-4281) during m3u8 playlist generation.

hlsGroupSettings_hlsCdnSettings :: Lens' HlsGroupSettings (Maybe HlsCdnSettings)

Parameters that control interactions with the CDN.

hlsGroupSettings_mode :: Lens' HlsGroupSettings (Maybe HlsMode)

If "vod", all segments are indexed and kept permanently in the destination and manifest. If "live", only the number segments specified in keepSegments and indexNSegments are kept; newer segments replace older segments, which may prevent players from rewinding all the way to the beginning of the event. VOD mode uses HLS EXT-X-PLAYLIST-TYPE of EVENT while the channel is running, converting it to a "VOD" type manifest on completion of the stream.

hlsGroupSettings_streamInfResolution :: Lens' HlsGroupSettings (Maybe HlsStreamInfResolution)

Include or exclude RESOLUTION attribute for video in EXT-X-STREAM-INF tag of variant manifest.

hlsGroupSettings_indexNSegments :: Lens' HlsGroupSettings (Maybe Natural)

Applies only if Mode field is LIVE. Specifies the maximum number of segments in the media manifest file. After this maximum, older segments are removed from the media manifest. This number must be smaller than the number in the Keep Segments field.

hlsGroupSettings_redundantManifest :: Lens' HlsGroupSettings (Maybe HlsRedundantManifest)

ENABLED: The master manifest (.m3u8 file) for each pipeline includes information about both pipelines: first its own media files, then the media files of the other pipeline. This feature allows playout device that support stale manifest detection to switch from one manifest to the other, when the current manifest seems to be stale. There are still two destinations and two master manifests, but both master manifests reference the media files from both pipelines. DISABLED: The master manifest (.m3u8 file) for each pipeline includes information about its own pipeline only. For an HLS output group with MediaPackage as the destination, the DISABLED behavior is always followed. MediaPackage regenerates the manifests it serves to players so a redundant manifest from MediaLive is irrelevant.

hlsGroupSettings_segmentationMode :: Lens' HlsGroupSettings (Maybe HlsSegmentationMode)

useInputSegmentation has been deprecated. The configured segment size is always used.

hlsGroupSettings_programDateTime :: Lens' HlsGroupSettings (Maybe HlsProgramDateTime)

Includes or excludes EXT-X-PROGRAM-DATE-TIME tag in .m3u8 manifest files. The value is calculated as follows: either the program date and time are initialized using the input timecode source, or the time is initialized using the input timecode source and the date is initialized using the timestampOffset.

hlsGroupSettings_iFrameOnlyPlaylists :: Lens' HlsGroupSettings (Maybe IFrameOnlyPlaylistType)

DISABLED: Do not create an I-frame-only manifest, but do create the master and media manifests (according to the Output Selection field). STANDARD: Create an I-frame-only manifest for each output that contains video, as well as the other manifests (according to the Output Selection field). The I-frame manifest contains a #EXT-X-I-FRAMES-ONLY tag to indicate it is I-frame only, and one or more #EXT-X-BYTERANGE entries identifying the I-frame position. For example, #EXT-X-BYTERANGE:160364@1461888"

hlsGroupSettings_tsFileMode :: Lens' HlsGroupSettings (Maybe HlsTsFileMode)

SEGMENTED_FILES: Emit the program as segments - multiple .ts media files. SINGLE_FILE: Applies only if Mode field is VOD. Emit the program as a single .ts media file. The media manifest includes #EXT-X-BYTERANGE tags to index segments for playback. A typical use for this value is when sending the output to AWS Elemental MediaConvert, which can accept only a single media file. Playback while the channel is running is not guaranteed due to HTTP server caching.

hlsGroupSettings_segmentsPerSubdirectory :: Lens' HlsGroupSettings (Maybe Natural)

Number of segments to write to a subdirectory before starting a new one. directoryStructure must be subdirectoryPerStream for this setting to have an effect.

hlsGroupSettings_discontinuityTags :: Lens' HlsGroupSettings (Maybe HlsDiscontinuityTags)

Specifies whether to insert EXT-X-DISCONTINUITY tags in the HLS child manifests for this output group. Typically, choose Insert because these tags are required in the manifest (according to the HLS specification) and serve an important purpose. Choose Never Insert only if the downstream system is doing real-time failover (without using the MediaLive automatic failover feature) and only if that downstream system has advised you to exclude the tags.

hlsGroupSettings_baseUrlContent :: Lens' HlsGroupSettings (Maybe Text)

A partial URI prefix that will be prepended to each output in the media .m3u8 file. Can be used if base manifest is delivered from a different URL than the main .m3u8 file.

hlsGroupSettings_hlsId3SegmentTagging :: Lens' HlsGroupSettings (Maybe HlsId3SegmentTaggingState)

State of HLS ID3 Segment Tagging

hlsGroupSettings_baseUrlManifest :: Lens' HlsGroupSettings (Maybe Text)

A partial URI prefix that will be prepended to each output in the media .m3u8 file. Can be used if base manifest is delivered from a different URL than the main .m3u8 file.

hlsGroupSettings_incompleteSegmentBehavior :: Lens' HlsGroupSettings (Maybe HlsIncompleteSegmentBehavior)

Specifies whether to include the final (incomplete) segment in the media output when the pipeline stops producing output because of a channel stop, a channel pause or a loss of input to the pipeline. Auto means that MediaLive decides whether to include the final segment, depending on the channel class and the types of output groups. Suppress means to never include the incomplete segment. We recommend you choose Auto and let MediaLive control the behavior.

hlsGroupSettings_baseUrlContent1 :: Lens' HlsGroupSettings (Maybe Text)

Optional. One value per output group. This field is required only if you are completing Base URL content A, and the downstream system has notified you that the media files for pipeline 1 of all outputs are in a location different from the media files for pipeline 0.

hlsGroupSettings_inputLossAction :: Lens' HlsGroupSettings (Maybe InputLossActionForHlsOut)

Parameter that control output group behavior on input loss.

hlsGroupSettings_keyProviderSettings :: Lens' HlsGroupSettings (Maybe KeyProviderSettings)

The key provider settings.

hlsGroupSettings_keepSegments :: Lens' HlsGroupSettings (Maybe Natural)

Applies only if Mode field is LIVE. Specifies the number of media segments to retain in the destination directory. This number should be bigger than indexNSegments (Num segments). We recommend (value = (2 x indexNsegments) + 1). If this "keep segments" number is too low, the following might happen: the player is still reading a media manifest file that lists this segment, but that segment has been removed from the destination directory (as directed by indexNSegments). This situation would result in a 404 HTTP error on the player.

hlsGroupSettings_captionLanguageMappings :: Lens' HlsGroupSettings (Maybe [CaptionLanguageMapping])

Mapping of up to 4 caption channels to caption languages. Is only meaningful if captionLanguageSetting is set to "insert".

hlsGroupSettings_timestampDeltaMilliseconds :: Lens' HlsGroupSettings (Maybe Natural)

Provides an extra millisecond delta offset to fine tune the timestamps.

hlsGroupSettings_baseUrlManifest1 :: Lens' HlsGroupSettings (Maybe Text)

Optional. One value per output group. Complete this field only if you are completing Base URL manifest A, and the downstream system has notified you that the child manifest files for pipeline 1 of all outputs are in a location different from the child manifest files for pipeline 0.

hlsGroupSettings_programDateTimePeriod :: Lens' HlsGroupSettings (Maybe Natural)

Period of insertion of EXT-X-PROGRAM-DATE-TIME entry, in seconds.

hlsGroupSettings_clientCache :: Lens' HlsGroupSettings (Maybe HlsClientCache)

When set to "disabled", sets the #EXT-X-ALLOW-CACHE:no tag in the manifest, which prevents clients from saving media segments for later replay.

hlsGroupSettings_keyFormatVersions :: Lens' HlsGroupSettings (Maybe Text)

Either a single positive integer version value or a slash delimited list of version values (1/2/3).

hlsGroupSettings_minSegmentLength :: Lens' HlsGroupSettings (Maybe Natural)

When set, minimumSegmentLength is enforced by looking ahead and back within the specified range for a nearby avail and extending the segment size if needed.

hlsGroupSettings_ivSource :: Lens' HlsGroupSettings (Maybe HlsIvSource)

For use with encryptionType. The IV (Initialization Vector) is a 128-bit number used in conjunction with the key for encrypting blocks. If this setting is "followsSegmentNumber", it will cause the IV to change every segment (to match the segment number). If this is set to "explicit", you must enter a constantIv value.

hlsGroupSettings_captionLanguageSetting :: Lens' HlsGroupSettings (Maybe HlsCaptionLanguageSetting)

Applies only to 608 Embedded output captions. insert: Include CLOSED-CAPTIONS lines in the manifest. Specify at least one language in the CC1 Language Code field. One CLOSED-CAPTION line is added for each Language Code you specify. Make sure to specify the languages in the order in which they appear in the original source (if the source is embedded format) or the order of the caption selectors (if the source is other than embedded). Otherwise, languages in the manifest will not match up properly with the output captions. none: Include CLOSED-CAPTIONS=NONE line in the manifest. omit: Omit any CLOSED-CAPTIONS line from the manifest.

hlsGroupSettings_manifestDurationFormat :: Lens' HlsGroupSettings (Maybe HlsManifestDurationFormat)

Indicates whether the output manifest should use floating point or integer values for segment duration.

hlsGroupSettings_destination :: Lens' HlsGroupSettings OutputLocationRef

A directory or HTTP destination for the HLS segments, manifest files, and encryption keys (if enabled).