Network.AWS.MediaLive.Types.HlsInputSettings

Description

Documentation

data HlsInputSettings

Hls Input Settings

See: newHlsInputSettings smart constructor.

Constructors

HlsInputSettings'

Fields

retryInterval :: Maybe Natural

The number of seconds between retries when an attempt to read a manifest or segment fails.

bandwidth :: Maybe Natural

When specified the HLS stream with the m3u8 BANDWIDTH that most closely matches this value will be chosen, otherwise the highest bandwidth stream in the m3u8 will be chosen. The bitrate is specified in bits per second, as in an HLS manifest.

retries :: Maybe Natural

The number of consecutive times that attempts to read a manifest or segment must fail before the input is considered unavailable.

bufferSegments :: Maybe Natural

When specified, reading of the HLS input will begin this many buffer segments from the end (most recently written segment). When not specified, the HLS input will begin with the first segment specified in the m3u8.

scte35Source :: Maybe HlsScte35SourceType

Identifies the source for the SCTE-35 messages that MediaLive will ingest. Messages can be ingested from the content segments (in the stream) or from tags in the playlist (the HLS manifest). MediaLive ignores SCTE-35 information in the source that is not selected.

newHlsInputSettings :: HlsInputSettings

Create a value of HlsInputSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:retryInterval:HlsInputSettings', hlsInputSettings_retryInterval - The number of seconds between retries when an attempt to read a manifest or segment fails.

$sel:bandwidth:HlsInputSettings', hlsInputSettings_bandwidth - When specified the HLS stream with the m3u8 BANDWIDTH that most closely matches this value will be chosen, otherwise the highest bandwidth stream in the m3u8 will be chosen. The bitrate is specified in bits per second, as in an HLS manifest.

$sel:retries:HlsInputSettings', hlsInputSettings_retries - The number of consecutive times that attempts to read a manifest or segment must fail before the input is considered unavailable.

$sel:bufferSegments:HlsInputSettings', hlsInputSettings_bufferSegments - When specified, reading of the HLS input will begin this many buffer segments from the end (most recently written segment). When not specified, the HLS input will begin with the first segment specified in the m3u8.

$sel:scte35Source:HlsInputSettings', hlsInputSettings_scte35Source - Identifies the source for the SCTE-35 messages that MediaLive will ingest. Messages can be ingested from the content segments (in the stream) or from tags in the playlist (the HLS manifest). MediaLive ignores SCTE-35 information in the source that is not selected.

hlsInputSettings_retryInterval :: Lens' HlsInputSettings (Maybe Natural)

The number of seconds between retries when an attempt to read a manifest or segment fails.

hlsInputSettings_bandwidth :: Lens' HlsInputSettings (Maybe Natural)

When specified the HLS stream with the m3u8 BANDWIDTH that most closely matches this value will be chosen, otherwise the highest bandwidth stream in the m3u8 will be chosen. The bitrate is specified in bits per second, as in an HLS manifest.

hlsInputSettings_retries :: Lens' HlsInputSettings (Maybe Natural)

The number of consecutive times that attempts to read a manifest or segment must fail before the input is considered unavailable.

hlsInputSettings_bufferSegments :: Lens' HlsInputSettings (Maybe Natural)

When specified, reading of the HLS input will begin this many buffer segments from the end (most recently written segment). When not specified, the HLS input will begin with the first segment specified in the m3u8.

hlsInputSettings_scte35Source :: Lens' HlsInputSettings (Maybe HlsScte35SourceType)

Identifies the source for the SCTE-35 messages that MediaLive will ingest. Messages can be ingested from the content segments (in the stream) or from tags in the playlist (the HLS manifest). MediaLive ignores SCTE-35 information in the source that is not selected.