Network.AWS.MediaLive.Types.HlsMediaStoreSettings

Description

Documentation

data HlsMediaStoreSettings

Hls Media Store Settings

See: newHlsMediaStoreSettings smart constructor.

Constructors

HlsMediaStoreSettings'

Fields

filecacheDuration :: Maybe Natural

Size in seconds of file cache for streaming outputs.

numRetries :: Maybe Natural

Number of retry attempts that will be made before the Live Event is put into an error state.

mediaStoreStorageClass :: Maybe HlsMediaStoreStorageClass

When set to temporal, output files are stored in non-persistent memory for faster reading and writing.

connectionRetryInterval :: Maybe Natural

Number of seconds to wait before retrying connection to the CDN if the connection is lost.

restartDelay :: Maybe Natural

If a streaming output fails, number of seconds to wait until a restart is initiated. A value of 0 means never restart.

newHlsMediaStoreSettings :: HlsMediaStoreSettings

Create a value of HlsMediaStoreSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filecacheDuration:HlsMediaStoreSettings', hlsMediaStoreSettings_filecacheDuration - Size in seconds of file cache for streaming outputs.

$sel:numRetries:HlsMediaStoreSettings', hlsMediaStoreSettings_numRetries - Number of retry attempts that will be made before the Live Event is put into an error state.

$sel:mediaStoreStorageClass:HlsMediaStoreSettings', hlsMediaStoreSettings_mediaStoreStorageClass - When set to temporal, output files are stored in non-persistent memory for faster reading and writing.

$sel:connectionRetryInterval:HlsMediaStoreSettings', hlsMediaStoreSettings_connectionRetryInterval - Number of seconds to wait before retrying connection to the CDN if the connection is lost.

$sel:restartDelay:HlsMediaStoreSettings', hlsMediaStoreSettings_restartDelay - If a streaming output fails, number of seconds to wait until a restart is initiated. A value of 0 means never restart.

hlsMediaStoreSettings_filecacheDuration :: Lens' HlsMediaStoreSettings (Maybe Natural)

Size in seconds of file cache for streaming outputs.

hlsMediaStoreSettings_numRetries :: Lens' HlsMediaStoreSettings (Maybe Natural)

Number of retry attempts that will be made before the Live Event is put into an error state.

hlsMediaStoreSettings_mediaStoreStorageClass :: Lens' HlsMediaStoreSettings (Maybe HlsMediaStoreStorageClass)

When set to temporal, output files are stored in non-persistent memory for faster reading and writing.

hlsMediaStoreSettings_connectionRetryInterval :: Lens' HlsMediaStoreSettings (Maybe Natural)

Number of seconds to wait before retrying connection to the CDN if the connection is lost.

hlsMediaStoreSettings_restartDelay :: Lens' HlsMediaStoreSettings (Maybe Natural)

If a streaming output fails, number of seconds to wait until a restart is initiated. A value of 0 means never restart.