Network.AWS.MediaLive.Types.Input

Description

Documentation

data Input

Placeholder documentation for Input

See: newInput smart constructor.

Constructors

Input'

Fields

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the role this input assumes during and after creation.

inputPartnerIds :: Maybe [Text]

A list of IDs for all Inputs which are partners of this one.

sources :: Maybe [InputSource]

A list of the sources of the input (PULL-type).

mediaConnectFlows :: Maybe [MediaConnectFlow]

A list of MediaConnect Flows for this input.

inputSourceType :: Maybe InputSourceType

Certain pull input sources can be dynamic, meaning that they can have their URL's dynamically changes during input switch actions. Presently, this functionality only works with MP4_FILE inputs.

arn :: Maybe Text

The Unique ARN of the input (generated, immutable).

id :: Maybe Text

The generated ID of the input (unique for user account, immutable).

securityGroups :: Maybe [Text]

A list of IDs for all the Input Security Groups attached to the input.

name :: Maybe Text

The user-assigned name (This is a mutable value).

destinations :: Maybe [InputDestination]

A list of the destinations of the input (PUSH-type).

state :: Maybe InputState

tags :: Maybe (HashMap Text Text)

A collection of key-value pairs.

inputClass :: Maybe InputClass

STANDARD - MediaLive expects two sources to be connected to this input. If the channel is also STANDARD, both sources will be ingested. If the channel is SINGLE_PIPELINE, only the first source will be ingested; the second source will always be ignored, even if the first source fails. SINGLE_PIPELINE - You can connect only one source to this input. If the ChannelClass is also SINGLE_PIPELINE, this value is valid. If the ChannelClass is STANDARD, this value is not valid because the channel requires two sources in the input.

type' :: Maybe InputType

inputDevices :: Maybe [InputDeviceSettings]

Settings for the input devices.

attachedChannels :: Maybe [Text]

A list of channel IDs that that input is attached to (currently an input can only be attached to one channel).

newInput :: Input

Create a value of Input with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleArn:Input', input_roleArn - The Amazon Resource Name (ARN) of the role this input assumes during and after creation.

$sel:inputPartnerIds:Input', input_inputPartnerIds - A list of IDs for all Inputs which are partners of this one.

$sel:sources:Input', input_sources - A list of the sources of the input (PULL-type).

$sel:mediaConnectFlows:Input', input_mediaConnectFlows - A list of MediaConnect Flows for this input.

$sel:inputSourceType:Input', input_inputSourceType - Certain pull input sources can be dynamic, meaning that they can have their URL's dynamically changes during input switch actions. Presently, this functionality only works with MP4_FILE inputs.

$sel:arn:Input', input_arn - The Unique ARN of the input (generated, immutable).

$sel:id:Input', input_id - The generated ID of the input (unique for user account, immutable).

$sel:securityGroups:Input', input_securityGroups - A list of IDs for all the Input Security Groups attached to the input.

$sel:name:Input', input_name - The user-assigned name (This is a mutable value).

$sel:destinations:Input', input_destinations - A list of the destinations of the input (PUSH-type).

$sel:state:Input', input_state - Undocumented member.

$sel:tags:Input', input_tags - A collection of key-value pairs.

$sel:inputClass:Input', input_inputClass - STANDARD - MediaLive expects two sources to be connected to this input. If the channel is also STANDARD, both sources will be ingested. If the channel is SINGLE_PIPELINE, only the first source will be ingested; the second source will always be ignored, even if the first source fails. SINGLE_PIPELINE - You can connect only one source to this input. If the ChannelClass is also SINGLE_PIPELINE, this value is valid. If the ChannelClass is STANDARD, this value is not valid because the channel requires two sources in the input.

$sel:type':Input', input_type - Undocumented member.

$sel:inputDevices:Input', input_inputDevices - Settings for the input devices.

$sel:attachedChannels:Input', input_attachedChannels - A list of channel IDs that that input is attached to (currently an input can only be attached to one channel).

input_roleArn :: Lens' Input (Maybe Text)

The Amazon Resource Name (ARN) of the role this input assumes during and after creation.

input_inputPartnerIds :: Lens' Input (Maybe [Text])

A list of IDs for all Inputs which are partners of this one.

input_sources :: Lens' Input (Maybe [InputSource])

A list of the sources of the input (PULL-type).

input_mediaConnectFlows :: Lens' Input (Maybe [MediaConnectFlow])

A list of MediaConnect Flows for this input.

input_inputSourceType :: Lens' Input (Maybe InputSourceType)

Certain pull input sources can be dynamic, meaning that they can have their URL's dynamically changes during input switch actions. Presently, this functionality only works with MP4_FILE inputs.

input_arn :: Lens' Input (Maybe Text)

The Unique ARN of the input (generated, immutable).

input_id :: Lens' Input (Maybe Text)

The generated ID of the input (unique for user account, immutable).

input_securityGroups :: Lens' Input (Maybe [Text])

A list of IDs for all the Input Security Groups attached to the input.

input_name :: Lens' Input (Maybe Text)

The user-assigned name (This is a mutable value).

input_destinations :: Lens' Input (Maybe [InputDestination])

A list of the destinations of the input (PUSH-type).

input_state :: Lens' Input (Maybe InputState)

Undocumented member.

input_tags :: Lens' Input (Maybe (HashMap Text Text))

A collection of key-value pairs.

input_inputClass :: Lens' Input (Maybe InputClass)

STANDARD - MediaLive expects two sources to be connected to this input. If the channel is also STANDARD, both sources will be ingested. If the channel is SINGLE_PIPELINE, only the first source will be ingested; the second source will always be ignored, even if the first source fails. SINGLE_PIPELINE - You can connect only one source to this input. If the ChannelClass is also SINGLE_PIPELINE, this value is valid. If the ChannelClass is STANDARD, this value is not valid because the channel requires two sources in the input.

input_type :: Lens' Input (Maybe InputType)

Undocumented member.

input_inputDevices :: Lens' Input (Maybe [InputDeviceSettings])

Settings for the input devices.

input_attachedChannels :: Lens' Input (Maybe [Text])

A list of channel IDs that that input is attached to (currently an input can only be attached to one channel).