Network.AWS.MediaLive.Types.InputDeviceHdSettings

Description

Documentation

data InputDeviceHdSettings

Settings that describe the active source from the input device, and the video characteristics of that source.

See: newInputDeviceHdSettings smart constructor.

Constructors

InputDeviceHdSettings'

Fields

height :: Maybe Int

The height of the video source, in pixels.

scanType :: Maybe InputDeviceScanType

The scan type of the video source.

width :: Maybe Int

The width of the video source, in pixels.

configuredInput :: Maybe InputDeviceConfiguredInput

The source at the input device that is currently active. You can specify this source.

framerate :: Maybe Double

The frame rate of the video source.

deviceState :: Maybe InputDeviceState

The state of the input device.

maxBitrate :: Maybe Int

The current maximum bitrate for ingesting this source, in bits per second. You can specify this maximum.

activeInput :: Maybe InputDeviceActiveInput

If you specified Auto as the configured input, specifies which of the sources is currently active (SDI or HDMI).

newInputDeviceHdSettings :: InputDeviceHdSettings

Create a value of InputDeviceHdSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:height:InputDeviceHdSettings', inputDeviceHdSettings_height - The height of the video source, in pixels.

$sel:scanType:InputDeviceHdSettings', inputDeviceHdSettings_scanType - The scan type of the video source.

$sel:width:InputDeviceHdSettings', inputDeviceHdSettings_width - The width of the video source, in pixels.

$sel:configuredInput:InputDeviceHdSettings', inputDeviceHdSettings_configuredInput - The source at the input device that is currently active. You can specify this source.

$sel:framerate:InputDeviceHdSettings', inputDeviceHdSettings_framerate - The frame rate of the video source.

$sel:deviceState:InputDeviceHdSettings', inputDeviceHdSettings_deviceState - The state of the input device.

$sel:maxBitrate:InputDeviceHdSettings', inputDeviceHdSettings_maxBitrate - The current maximum bitrate for ingesting this source, in bits per second. You can specify this maximum.

$sel:activeInput:InputDeviceHdSettings', inputDeviceHdSettings_activeInput - If you specified Auto as the configured input, specifies which of the sources is currently active (SDI or HDMI).

inputDeviceHdSettings_height :: Lens' InputDeviceHdSettings (Maybe Int)

The height of the video source, in pixels.

inputDeviceHdSettings_scanType :: Lens' InputDeviceHdSettings (Maybe InputDeviceScanType)

The scan type of the video source.

inputDeviceHdSettings_width :: Lens' InputDeviceHdSettings (Maybe Int)

The width of the video source, in pixels.

inputDeviceHdSettings_configuredInput :: Lens' InputDeviceHdSettings (Maybe InputDeviceConfiguredInput)

The source at the input device that is currently active. You can specify this source.

inputDeviceHdSettings_framerate :: Lens' InputDeviceHdSettings (Maybe Double)

The frame rate of the video source.

inputDeviceHdSettings_deviceState :: Lens' InputDeviceHdSettings (Maybe InputDeviceState)

The state of the input device.

inputDeviceHdSettings_maxBitrate :: Lens' InputDeviceHdSettings (Maybe Int)

The current maximum bitrate for ingesting this source, in bits per second. You can specify this maximum.

inputDeviceHdSettings_activeInput :: Lens' InputDeviceHdSettings (Maybe InputDeviceActiveInput)

If you specified Auto as the configured input, specifies which of the sources is currently active (SDI or HDMI).