Network.AWS.MediaLive.Types.InputDeviceUhdSettings

Description

Documentation

data InputDeviceUhdSettings

Settings that describe the active source from the input device, and the video characteristics of that source.

See: newInputDeviceUhdSettings smart constructor.

Constructors

InputDeviceUhdSettings'

Fields

height :: Maybe Int

The height of the video source, in pixels.

scanType :: Maybe InputDeviceScanType

The scan type of the video source.

width :: Maybe Int

The width of the video source, in pixels.

configuredInput :: Maybe InputDeviceConfiguredInput

The source at the input device that is currently active. You can specify this source.

framerate :: Maybe Double

The frame rate of the video source.

deviceState :: Maybe InputDeviceState

The state of the input device.

maxBitrate :: Maybe Int

The current maximum bitrate for ingesting this source, in bits per second. You can specify this maximum.

activeInput :: Maybe InputDeviceActiveInput

If you specified Auto as the configured input, specifies which of the sources is currently active (SDI or HDMI).

newInputDeviceUhdSettings :: InputDeviceUhdSettings

Create a value of InputDeviceUhdSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:height:InputDeviceUhdSettings', inputDeviceUhdSettings_height - The height of the video source, in pixels.

$sel:scanType:InputDeviceUhdSettings', inputDeviceUhdSettings_scanType - The scan type of the video source.

$sel:width:InputDeviceUhdSettings', inputDeviceUhdSettings_width - The width of the video source, in pixels.

$sel:configuredInput:InputDeviceUhdSettings', inputDeviceUhdSettings_configuredInput - The source at the input device that is currently active. You can specify this source.

$sel:framerate:InputDeviceUhdSettings', inputDeviceUhdSettings_framerate - The frame rate of the video source.

$sel:deviceState:InputDeviceUhdSettings', inputDeviceUhdSettings_deviceState - The state of the input device.

$sel:maxBitrate:InputDeviceUhdSettings', inputDeviceUhdSettings_maxBitrate - The current maximum bitrate for ingesting this source, in bits per second. You can specify this maximum.

$sel:activeInput:InputDeviceUhdSettings', inputDeviceUhdSettings_activeInput - If you specified Auto as the configured input, specifies which of the sources is currently active (SDI or HDMI).

inputDeviceUhdSettings_height :: Lens' InputDeviceUhdSettings (Maybe Int)

The height of the video source, in pixels.

inputDeviceUhdSettings_scanType :: Lens' InputDeviceUhdSettings (Maybe InputDeviceScanType)

The scan type of the video source.

inputDeviceUhdSettings_width :: Lens' InputDeviceUhdSettings (Maybe Int)

The width of the video source, in pixels.

inputDeviceUhdSettings_configuredInput :: Lens' InputDeviceUhdSettings (Maybe InputDeviceConfiguredInput)

The source at the input device that is currently active. You can specify this source.

inputDeviceUhdSettings_framerate :: Lens' InputDeviceUhdSettings (Maybe Double)

The frame rate of the video source.

inputDeviceUhdSettings_deviceState :: Lens' InputDeviceUhdSettings (Maybe InputDeviceState)

The state of the input device.

inputDeviceUhdSettings_maxBitrate :: Lens' InputDeviceUhdSettings (Maybe Int)

The current maximum bitrate for ingesting this source, in bits per second. You can specify this maximum.

inputDeviceUhdSettings_activeInput :: Lens' InputDeviceUhdSettings (Maybe InputDeviceActiveInput)

If you specified Auto as the configured input, specifies which of the sources is currently active (SDI or HDMI).