Network.AWS.MediaLive.Types.InputLocation

Description

Documentation

data InputLocation

Input Location

See: newInputLocation smart constructor.

Constructors

InputLocation'

Fields

passwordParam :: Maybe Text

key used to extract the password from EC2 Parameter store

username :: Maybe Text

Documentation update needed

uri :: Text

Uniform Resource Identifier - This should be a path to a file accessible to the Live system (eg. a http:// URI) depending on the output type. For example, a RTMP destination should have a uri simliar to: "rtmp://fmsserver/live".

newInputLocation

Arguments

:: Text
$sel:uri:InputLocation'

-> InputLocation

Create a value of InputLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:passwordParam:InputLocation', inputLocation_passwordParam - key used to extract the password from EC2 Parameter store

$sel:username:InputLocation', inputLocation_username - Documentation update needed

$sel:uri:InputLocation', inputLocation_uri - Uniform Resource Identifier - This should be a path to a file accessible to the Live system (eg. a http:// URI) depending on the output type. For example, a RTMP destination should have a uri simliar to: "rtmp://fmsserver/live".

inputLocation_passwordParam :: Lens' InputLocation (Maybe Text)

key used to extract the password from EC2 Parameter store

inputLocation_username :: Lens' InputLocation (Maybe Text)

Documentation update needed

inputLocation_uri :: Lens' InputLocation Text

Uniform Resource Identifier - This should be a path to a file accessible to the Live system (eg. a http:// URI) depending on the output type. For example, a RTMP destination should have a uri simliar to: "rtmp://fmsserver/live".