Network.AWS.MediaLive.Types.InputLossBehavior

Description

Documentation

data InputLossBehavior

Input Loss Behavior

See: newInputLossBehavior smart constructor.

Constructors

InputLossBehavior'

Fields

blackFrameMsec :: Maybe Natural

Documentation update needed

inputLossImageColor :: Maybe Text

When input loss image type is "color" this field specifies the color to use. Value: 6 hex characters representing the values of RGB.

inputLossImageSlate :: Maybe InputLocation

When input loss image type is "slate" these fields specify the parameters for accessing the slate.

repeatFrameMsec :: Maybe Natural

Documentation update needed

inputLossImageType :: Maybe InputLossImageType

Indicates whether to substitute a solid color or a slate into the output after input loss exceeds blackFrameMsec.

newInputLossBehavior :: InputLossBehavior

Create a value of InputLossBehavior with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blackFrameMsec:InputLossBehavior', inputLossBehavior_blackFrameMsec - Documentation update needed

$sel:inputLossImageColor:InputLossBehavior', inputLossBehavior_inputLossImageColor - When input loss image type is "color" this field specifies the color to use. Value: 6 hex characters representing the values of RGB.

$sel:inputLossImageSlate:InputLossBehavior', inputLossBehavior_inputLossImageSlate - When input loss image type is "slate" these fields specify the parameters for accessing the slate.

$sel:repeatFrameMsec:InputLossBehavior', inputLossBehavior_repeatFrameMsec - Documentation update needed

$sel:inputLossImageType:InputLossBehavior', inputLossBehavior_inputLossImageType - Indicates whether to substitute a solid color or a slate into the output after input loss exceeds blackFrameMsec.

inputLossBehavior_blackFrameMsec :: Lens' InputLossBehavior (Maybe Natural)

Documentation update needed

inputLossBehavior_inputLossImageColor :: Lens' InputLossBehavior (Maybe Text)

When input loss image type is "color" this field specifies the color to use. Value: 6 hex characters representing the values of RGB.

inputLossBehavior_inputLossImageSlate :: Lens' InputLossBehavior (Maybe InputLocation)

When input loss image type is "slate" these fields specify the parameters for accessing the slate.

inputLossBehavior_repeatFrameMsec :: Lens' InputLossBehavior (Maybe Natural)

Documentation update needed

inputLossBehavior_inputLossImageType :: Lens' InputLossBehavior (Maybe InputLossImageType)

Indicates whether to substitute a solid color or a slate into the output after input loss exceeds blackFrameMsec.