Network.AWS.MediaLive.Types.InputSettings

Description

Documentation

data InputSettings

Live Event input parameters. There can be multiple inputs in a single Live Event.

See: newInputSettings smart constructor.

Constructors

InputSettings'

Fields

videoSelector :: Maybe VideoSelector

Informs which video elementary stream to decode for input types that have multiple available.

smpte2038DataPreference :: Maybe Smpte2038DataPreference

Specifies whether to extract applicable ancillary data from a SMPTE-2038 source in this input. Applicable data types are captions, timecode, AFD, and SCTE-104 messages. - PREFER: Extract from SMPTE-2038 if present in this input, otherwise extract from another source (if any). - IGNORE: Never extract any ancillary data from SMPTE-2038.

networkInputSettings :: Maybe NetworkInputSettings

Input settings.

audioSelectors :: Maybe [AudioSelector]

Used to select the audio stream to decode for inputs that have multiple available.

deblockFilter :: Maybe InputDeblockFilter

Enable or disable the deblock filter when filtering.

denoiseFilter :: Maybe InputDenoiseFilter

Enable or disable the denoise filter when filtering.

filterStrength :: Maybe Natural

Adjusts the magnitude of filtering from 1 (minimal) to 5 (strongest).

captionSelectors :: Maybe [CaptionSelector]

Used to select the caption input to use for inputs that have multiple available.

inputFilter :: Maybe InputFilter

Turns on the filter for this input. MPEG-2 inputs have the deblocking filter enabled by default. 1) auto - filtering will be applied depending on input type/quality 2) disabled - no filtering will be applied to the input 3) forced - filtering will be applied regardless of input type

sourceEndBehavior :: Maybe InputSourceEndBehavior

Loop input if it is a file. This allows a file input to be streamed indefinitely.

newInputSettings :: InputSettings

Create a value of InputSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:videoSelector:InputSettings', inputSettings_videoSelector - Informs which video elementary stream to decode for input types that have multiple available.

$sel:smpte2038DataPreference:InputSettings', inputSettings_smpte2038DataPreference - Specifies whether to extract applicable ancillary data from a SMPTE-2038 source in this input. Applicable data types are captions, timecode, AFD, and SCTE-104 messages. - PREFER: Extract from SMPTE-2038 if present in this input, otherwise extract from another source (if any). - IGNORE: Never extract any ancillary data from SMPTE-2038.

$sel:networkInputSettings:InputSettings', inputSettings_networkInputSettings - Input settings.

$sel:audioSelectors:InputSettings', inputSettings_audioSelectors - Used to select the audio stream to decode for inputs that have multiple available.

$sel:deblockFilter:InputSettings', inputSettings_deblockFilter - Enable or disable the deblock filter when filtering.

$sel:denoiseFilter:InputSettings', inputSettings_denoiseFilter - Enable or disable the denoise filter when filtering.

$sel:filterStrength:InputSettings', inputSettings_filterStrength - Adjusts the magnitude of filtering from 1 (minimal) to 5 (strongest).

$sel:captionSelectors:InputSettings', inputSettings_captionSelectors - Used to select the caption input to use for inputs that have multiple available.

$sel:inputFilter:InputSettings', inputSettings_inputFilter - Turns on the filter for this input. MPEG-2 inputs have the deblocking filter enabled by default. 1) auto - filtering will be applied depending on input type/quality 2) disabled - no filtering will be applied to the input 3) forced - filtering will be applied regardless of input type

$sel:sourceEndBehavior:InputSettings', inputSettings_sourceEndBehavior - Loop input if it is a file. This allows a file input to be streamed indefinitely.

inputSettings_videoSelector :: Lens' InputSettings (Maybe VideoSelector)

Informs which video elementary stream to decode for input types that have multiple available.

inputSettings_smpte2038DataPreference :: Lens' InputSettings (Maybe Smpte2038DataPreference)

Specifies whether to extract applicable ancillary data from a SMPTE-2038 source in this input. Applicable data types are captions, timecode, AFD, and SCTE-104 messages. - PREFER: Extract from SMPTE-2038 if present in this input, otherwise extract from another source (if any). - IGNORE: Never extract any ancillary data from SMPTE-2038.

inputSettings_networkInputSettings :: Lens' InputSettings (Maybe NetworkInputSettings)

Input settings.

inputSettings_audioSelectors :: Lens' InputSettings (Maybe [AudioSelector])

Used to select the audio stream to decode for inputs that have multiple available.

inputSettings_deblockFilter :: Lens' InputSettings (Maybe InputDeblockFilter)

Enable or disable the deblock filter when filtering.

inputSettings_denoiseFilter :: Lens' InputSettings (Maybe InputDenoiseFilter)

Enable or disable the denoise filter when filtering.

inputSettings_filterStrength :: Lens' InputSettings (Maybe Natural)

Adjusts the magnitude of filtering from 1 (minimal) to 5 (strongest).

inputSettings_captionSelectors :: Lens' InputSettings (Maybe [CaptionSelector])

Used to select the caption input to use for inputs that have multiple available.

inputSettings_inputFilter :: Lens' InputSettings (Maybe InputFilter)

Turns on the filter for this input. MPEG-2 inputs have the deblocking filter enabled by default. 1) auto - filtering will be applied depending on input type/quality 2) disabled - no filtering will be applied to the input 3) forced - filtering will be applied regardless of input type

inputSettings_sourceEndBehavior :: Lens' InputSettings (Maybe InputSourceEndBehavior)

Loop input if it is a file. This allows a file input to be streamed indefinitely.