Network.AWS.MediaLive.Types.InputSwitchScheduleActionSettings

Description

Documentation

data InputSwitchScheduleActionSettings

Settings for the "switch input" action: to switch from ingesting one input to ingesting another input.

See: newInputSwitchScheduleActionSettings smart constructor.

Constructors

InputSwitchScheduleActionSettings'

Fields

urlPath :: Maybe [Text]

The value for the variable portion of the URL for the dynamic input, for this instance of the input. Each time you use the same dynamic input in an input switch action, you can provide a different value, in order to connect the input to a different content source.

inputClippingSettings :: Maybe InputClippingSettings

Settings to let you create a clip of the file input, in order to set up the input to ingest only a portion of the file.

inputAttachmentNameReference :: Text

The name of the input attachment (not the name of the input!) to switch to. The name is specified in the channel configuration.

newInputSwitchScheduleActionSettings

Arguments

:: Text
$sel:inputAttachmentNameReference:InputSwitchScheduleActionSettings'

-> InputSwitchScheduleActionSettings

Create a value of InputSwitchScheduleActionSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:urlPath:InputSwitchScheduleActionSettings', inputSwitchScheduleActionSettings_urlPath - The value for the variable portion of the URL for the dynamic input, for this instance of the input. Each time you use the same dynamic input in an input switch action, you can provide a different value, in order to connect the input to a different content source.

$sel:inputClippingSettings:InputSwitchScheduleActionSettings', inputSwitchScheduleActionSettings_inputClippingSettings - Settings to let you create a clip of the file input, in order to set up the input to ingest only a portion of the file.

$sel:inputAttachmentNameReference:InputSwitchScheduleActionSettings', inputSwitchScheduleActionSettings_inputAttachmentNameReference - The name of the input attachment (not the name of the input!) to switch to. The name is specified in the channel configuration.

inputSwitchScheduleActionSettings_urlPath :: Lens' InputSwitchScheduleActionSettings (Maybe [Text])

The value for the variable portion of the URL for the dynamic input, for this instance of the input. Each time you use the same dynamic input in an input switch action, you can provide a different value, in order to connect the input to a different content source.

inputSwitchScheduleActionSettings_inputClippingSettings :: Lens' InputSwitchScheduleActionSettings (Maybe InputClippingSettings)

Settings to let you create a clip of the file input, in order to set up the input to ingest only a portion of the file.

inputSwitchScheduleActionSettings_inputAttachmentNameReference :: Lens' InputSwitchScheduleActionSettings Text

The name of the input attachment (not the name of the input!) to switch to. The name is specified in the channel configuration.