Network.AWS.MediaLive.Types.Mpeg2Settings

Description

Documentation

data Mpeg2Settings

Mpeg2 Settings

See: newMpeg2Settings smart constructor.

Constructors

Mpeg2Settings'

Fields

scanType :: Maybe Mpeg2ScanType

Set the scan type of the output to PROGRESSIVE or INTERLACED (top field first).

displayAspectRatio :: Maybe Mpeg2DisplayRatio

Sets the pixel aspect ratio for the encode.

colorMetadata :: Maybe Mpeg2ColorMetadata

Specifies whether to include the color space metadata. The metadata describes the color space that applies to the video (the colorSpace field). We recommend that you insert the metadata.

gopSizeUnits :: Maybe Mpeg2GopSizeUnits

Relates to the GOP structure. Specifies whether the gopSize is specified in frames or seconds. If you do not plan to change the default gopSize, leave the default. If you specify SECONDS, MediaLive will internally convert the gop size to a frame count.

subgopLength :: Maybe Mpeg2SubGopLength

Relates to the GOP structure. If you do not know what GOP is, use the default. FIXED: Set the number of B-frames in each sub-GOP to the value in gopNumBFrames. DYNAMIC: Let MediaLive optimize the number of B-frames in each sub-GOP, to improve visual quality.

fixedAfd :: Maybe FixedAfd

Complete this field only when afdSignaling is set to FIXED. Enter the AFD value (4 bits) to write on all frames of the video encode.

gopSize :: Maybe Double

Relates to the GOP structure. The GOP size (keyframe interval) in the units specified in gopSizeUnits. If you do not know what GOP is, use the default. If gopSizeUnits is frames, then the gopSize must be an integer and must be greater than or equal to 1. If gopSizeUnits is seconds, the gopSize must be greater than 0, but does not need to be an integer.

filterSettings :: Maybe Mpeg2FilterSettings

Optionally specify a noise reduction filter, which can improve quality of compressed content. If you do not choose a filter, no filter will be applied. TEMPORAL: This filter is useful for both source content that is noisy (when it has excessive digital artifacts) and source content that is clean. When the content is noisy, the filter cleans up the source content before the encoding phase, with these two effects: First, it improves the output video quality because the content has been cleaned up. Secondly, it decreases the bandwidth because MediaLive does not waste bits on encoding noise. When the content is reasonably clean, the filter tends to decrease the bitrate.

gopNumBFrames :: Maybe Natural

Relates to the GOP structure. The number of B-frames between reference frames. If you do not know what a B-frame is, use the default.

timecodeInsertion :: Maybe Mpeg2TimecodeInsertionBehavior

Determines how MediaLive inserts timecodes in the output video. For detailed information about setting up the input and the output for a timecode, see the section on \"MediaLive Features - Timecode configuration\" in the MediaLive User Guide. DISABLED: do not include timecodes. GOP_TIMECODE: Include timecode metadata in the GOP header.

gopClosedCadence :: Maybe Natural

MPEG2: default is open GOP.

colorSpace :: Maybe Mpeg2ColorSpace

Choose the type of color space conversion to apply to the output. For detailed information on setting up both the input and the output to obtain the desired color space in the output, see the section on \"MediaLive Features - Video - color space\" in the MediaLive User Guide. PASSTHROUGH: Keep the color space of the input content - do not convert it. AUTO:Convert all content that is SD to rec 601, and convert all content that is HD to rec 709.

adaptiveQuantization :: Maybe Mpeg2AdaptiveQuantization

Choose Off to disable adaptive quantization. Or choose another value to enable the quantizer and set its strength. The strengths are: Auto, Off, Low, Medium, High. When you enable this field, MediaLive allows intra-frame quantizers to vary, which might improve visual quality.

afdSignaling :: Maybe AfdSignaling

Indicates the AFD values that MediaLive will write into the video encode. If you do not know what AFD signaling is, or if your downstream system has not given you guidance, choose AUTO. AUTO: MediaLive will try to preserve the input AFD value (in cases where multiple AFD values are valid). FIXED: MediaLive will use the value you specify in fixedAFD.

framerateNumerator :: Natural

The framerate numerator. For example, 24000. The framerate is the numerator divided by the denominator. For example, 24000 / 1001 = 23.976 FPS.

framerateDenominator :: Natural

description": "The framerate denominator. For example, 1001. The framerate is the numerator divided by the denominator. For example, 24000 / 1001 = 23.976 FPS.

newMpeg2Settings

Arguments

:: Natural
$sel:framerateNumerator:Mpeg2Settings'

-> Natural
$sel:framerateDenominator:Mpeg2Settings'

-> Mpeg2Settings

Create a value of Mpeg2Settings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scanType:Mpeg2Settings', mpeg2Settings_scanType - Set the scan type of the output to PROGRESSIVE or INTERLACED (top field first).

$sel:displayAspectRatio:Mpeg2Settings', mpeg2Settings_displayAspectRatio - Sets the pixel aspect ratio for the encode.

$sel:colorMetadata:Mpeg2Settings', mpeg2Settings_colorMetadata - Specifies whether to include the color space metadata. The metadata describes the color space that applies to the video (the colorSpace field). We recommend that you insert the metadata.

$sel:gopSizeUnits:Mpeg2Settings', mpeg2Settings_gopSizeUnits - Relates to the GOP structure. Specifies whether the gopSize is specified in frames or seconds. If you do not plan to change the default gopSize, leave the default. If you specify SECONDS, MediaLive will internally convert the gop size to a frame count.

$sel:subgopLength:Mpeg2Settings', mpeg2Settings_subgopLength - Relates to the GOP structure. If you do not know what GOP is, use the default. FIXED: Set the number of B-frames in each sub-GOP to the value in gopNumBFrames. DYNAMIC: Let MediaLive optimize the number of B-frames in each sub-GOP, to improve visual quality.

$sel:fixedAfd:Mpeg2Settings', mpeg2Settings_fixedAfd - Complete this field only when afdSignaling is set to FIXED. Enter the AFD value (4 bits) to write on all frames of the video encode.

$sel:gopSize:Mpeg2Settings', mpeg2Settings_gopSize - Relates to the GOP structure. The GOP size (keyframe interval) in the units specified in gopSizeUnits. If you do not know what GOP is, use the default. If gopSizeUnits is frames, then the gopSize must be an integer and must be greater than or equal to 1. If gopSizeUnits is seconds, the gopSize must be greater than 0, but does not need to be an integer.

$sel:filterSettings:Mpeg2Settings', mpeg2Settings_filterSettings - Optionally specify a noise reduction filter, which can improve quality of compressed content. If you do not choose a filter, no filter will be applied. TEMPORAL: This filter is useful for both source content that is noisy (when it has excessive digital artifacts) and source content that is clean. When the content is noisy, the filter cleans up the source content before the encoding phase, with these two effects: First, it improves the output video quality because the content has been cleaned up. Secondly, it decreases the bandwidth because MediaLive does not waste bits on encoding noise. When the content is reasonably clean, the filter tends to decrease the bitrate.

$sel:gopNumBFrames:Mpeg2Settings', mpeg2Settings_gopNumBFrames - Relates to the GOP structure. The number of B-frames between reference frames. If you do not know what a B-frame is, use the default.

$sel:timecodeInsertion:Mpeg2Settings', mpeg2Settings_timecodeInsertion - Determines how MediaLive inserts timecodes in the output video. For detailed information about setting up the input and the output for a timecode, see the section on \"MediaLive Features - Timecode configuration\" in the MediaLive User Guide. DISABLED: do not include timecodes. GOP_TIMECODE: Include timecode metadata in the GOP header.

$sel:gopClosedCadence:Mpeg2Settings', mpeg2Settings_gopClosedCadence - MPEG2: default is open GOP.

$sel:colorSpace:Mpeg2Settings', mpeg2Settings_colorSpace - Choose the type of color space conversion to apply to the output. For detailed information on setting up both the input and the output to obtain the desired color space in the output, see the section on \"MediaLive Features - Video - color space\" in the MediaLive User Guide. PASSTHROUGH: Keep the color space of the input content - do not convert it. AUTO:Convert all content that is SD to rec 601, and convert all content that is HD to rec 709.

$sel:adaptiveQuantization:Mpeg2Settings', mpeg2Settings_adaptiveQuantization - Choose Off to disable adaptive quantization. Or choose another value to enable the quantizer and set its strength. The strengths are: Auto, Off, Low, Medium, High. When you enable this field, MediaLive allows intra-frame quantizers to vary, which might improve visual quality.

$sel:afdSignaling:Mpeg2Settings', mpeg2Settings_afdSignaling - Indicates the AFD values that MediaLive will write into the video encode. If you do not know what AFD signaling is, or if your downstream system has not given you guidance, choose AUTO. AUTO: MediaLive will try to preserve the input AFD value (in cases where multiple AFD values are valid). FIXED: MediaLive will use the value you specify in fixedAFD.

$sel:framerateNumerator:Mpeg2Settings', mpeg2Settings_framerateNumerator - The framerate numerator. For example, 24000. The framerate is the numerator divided by the denominator. For example, 24000 / 1001 = 23.976 FPS.

$sel:framerateDenominator:Mpeg2Settings', mpeg2Settings_framerateDenominator - description": "The framerate denominator. For example, 1001. The framerate is the numerator divided by the denominator. For example, 24000 / 1001 = 23.976 FPS.

mpeg2Settings_scanType :: Lens' Mpeg2Settings (Maybe Mpeg2ScanType)

Set the scan type of the output to PROGRESSIVE or INTERLACED (top field first).

mpeg2Settings_displayAspectRatio :: Lens' Mpeg2Settings (Maybe Mpeg2DisplayRatio)

Sets the pixel aspect ratio for the encode.

mpeg2Settings_colorMetadata :: Lens' Mpeg2Settings (Maybe Mpeg2ColorMetadata)

Specifies whether to include the color space metadata. The metadata describes the color space that applies to the video (the colorSpace field). We recommend that you insert the metadata.

mpeg2Settings_gopSizeUnits :: Lens' Mpeg2Settings (Maybe Mpeg2GopSizeUnits)

Relates to the GOP structure. Specifies whether the gopSize is specified in frames or seconds. If you do not plan to change the default gopSize, leave the default. If you specify SECONDS, MediaLive will internally convert the gop size to a frame count.

mpeg2Settings_subgopLength :: Lens' Mpeg2Settings (Maybe Mpeg2SubGopLength)

Relates to the GOP structure. If you do not know what GOP is, use the default. FIXED: Set the number of B-frames in each sub-GOP to the value in gopNumBFrames. DYNAMIC: Let MediaLive optimize the number of B-frames in each sub-GOP, to improve visual quality.

mpeg2Settings_fixedAfd :: Lens' Mpeg2Settings (Maybe FixedAfd)

Complete this field only when afdSignaling is set to FIXED. Enter the AFD value (4 bits) to write on all frames of the video encode.

mpeg2Settings_gopSize :: Lens' Mpeg2Settings (Maybe Double)

Relates to the GOP structure. The GOP size (keyframe interval) in the units specified in gopSizeUnits. If you do not know what GOP is, use the default. If gopSizeUnits is frames, then the gopSize must be an integer and must be greater than or equal to 1. If gopSizeUnits is seconds, the gopSize must be greater than 0, but does not need to be an integer.

mpeg2Settings_filterSettings :: Lens' Mpeg2Settings (Maybe Mpeg2FilterSettings)

Optionally specify a noise reduction filter, which can improve quality of compressed content. If you do not choose a filter, no filter will be applied. TEMPORAL: This filter is useful for both source content that is noisy (when it has excessive digital artifacts) and source content that is clean. When the content is noisy, the filter cleans up the source content before the encoding phase, with these two effects: First, it improves the output video quality because the content has been cleaned up. Secondly, it decreases the bandwidth because MediaLive does not waste bits on encoding noise. When the content is reasonably clean, the filter tends to decrease the bitrate.

mpeg2Settings_gopNumBFrames :: Lens' Mpeg2Settings (Maybe Natural)

Relates to the GOP structure. The number of B-frames between reference frames. If you do not know what a B-frame is, use the default.

mpeg2Settings_timecodeInsertion :: Lens' Mpeg2Settings (Maybe Mpeg2TimecodeInsertionBehavior)

Determines how MediaLive inserts timecodes in the output video. For detailed information about setting up the input and the output for a timecode, see the section on \"MediaLive Features - Timecode configuration\" in the MediaLive User Guide. DISABLED: do not include timecodes. GOP_TIMECODE: Include timecode metadata in the GOP header.

mpeg2Settings_gopClosedCadence :: Lens' Mpeg2Settings (Maybe Natural)

MPEG2: default is open GOP.

mpeg2Settings_colorSpace :: Lens' Mpeg2Settings (Maybe Mpeg2ColorSpace)

Choose the type of color space conversion to apply to the output. For detailed information on setting up both the input and the output to obtain the desired color space in the output, see the section on \"MediaLive Features - Video - color space\" in the MediaLive User Guide. PASSTHROUGH: Keep the color space of the input content - do not convert it. AUTO:Convert all content that is SD to rec 601, and convert all content that is HD to rec 709.

mpeg2Settings_adaptiveQuantization :: Lens' Mpeg2Settings (Maybe Mpeg2AdaptiveQuantization)

Choose Off to disable adaptive quantization. Or choose another value to enable the quantizer and set its strength. The strengths are: Auto, Off, Low, Medium, High. When you enable this field, MediaLive allows intra-frame quantizers to vary, which might improve visual quality.

mpeg2Settings_afdSignaling :: Lens' Mpeg2Settings (Maybe AfdSignaling)

Indicates the AFD values that MediaLive will write into the video encode. If you do not know what AFD signaling is, or if your downstream system has not given you guidance, choose AUTO. AUTO: MediaLive will try to preserve the input AFD value (in cases where multiple AFD values are valid). FIXED: MediaLive will use the value you specify in fixedAFD.

mpeg2Settings_framerateNumerator :: Lens' Mpeg2Settings Natural

The framerate numerator. For example, 24000. The framerate is the numerator divided by the denominator. For example, 24000 / 1001 = 23.976 FPS.

mpeg2Settings_framerateDenominator :: Lens' Mpeg2Settings Natural

description": "The framerate denominator. For example, 1001. The framerate is the numerator divided by the denominator. For example, 24000 / 1001 = 23.976 FPS.