Network.AWS.MediaLive.Types.MsSmoothGroupSettings

Description

Documentation

data MsSmoothGroupSettings

Ms Smooth Group Settings

See: newMsSmoothGroupSettings smart constructor.

Constructors

MsSmoothGroupSettings'

Fields

streamManifestBehavior :: Maybe SmoothGroupStreamManifestBehavior

When set to send, send stream manifest so publishing point doesn't start until all streams start.

filecacheDuration :: Maybe Natural

Size in seconds of file cache for streaming outputs.

fragmentLength :: Maybe Natural

Length of mp4 fragments to generate (in seconds). Fragment length must be compatible with GOP size and framerate.

eventId :: Maybe Text

MS Smooth event ID to be sent to the IIS server. Should only be specified if eventIdMode is set to useConfigured.

certificateMode :: Maybe SmoothGroupCertificateMode

If set to verifyAuthenticity, verify the https certificate chain to a trusted Certificate Authority (CA). This will cause https outputs to self-signed certificates to fail.

acquisitionPointId :: Maybe Text

The ID to include in each message in the sparse track. Ignored if sparseTrackType is NONE.

numRetries :: Maybe Natural

Number of retry attempts.

segmentationMode :: Maybe SmoothGroupSegmentationMode

useInputSegmentation has been deprecated. The configured segment size is always used.

audioOnlyTimecodeControl :: Maybe SmoothGroupAudioOnlyTimecodeControl

If set to passthrough for an audio-only MS Smooth output, the fragment absolute time will be set to the current timecode. This option does not write timecodes to the audio elementary stream.

eventIdMode :: Maybe SmoothGroupEventIdMode

Specifies whether or not to send an event ID to the IIS server. If no event ID is sent and the same Live Event is used without changing the publishing point, clients might see cached video from the previous run. Options: - "useConfigured" - use the value provided in eventId - "useTimestamp" - generate and send an event ID based on the current timestamp - "noEventId" - do not send an event ID to the IIS server.

sendDelayMs :: Maybe Natural

Number of milliseconds to delay the output from the second pipeline.

sparseTrackType :: Maybe SmoothGroupSparseTrackType

Identifies the type of data to place in the sparse track: - SCTE35: Insert SCTE-35 messages from the source content. With each message, insert an IDR frame to start a new segment. - SCTE35_WITHOUT_SEGMENTATION: Insert SCTE-35 messages from the source content. With each message, insert an IDR frame but don't start a new segment. - NONE: Don't generate a sparse track for any outputs in this output group.

connectionRetryInterval :: Maybe Natural

Number of seconds to wait before retrying connection to the IIS server if the connection is lost. Content will be cached during this time and the cache will be be delivered to the IIS server once the connection is re-established.

inputLossAction :: Maybe InputLossActionForMsSmoothOut

Parameter that control output group behavior on input loss.

timestampOffset :: Maybe Text

Timestamp offset for the event. Only used if timestampOffsetMode is set to useConfiguredOffset.

eventStopBehavior :: Maybe SmoothGroupEventStopBehavior

When set to sendEos, send EOS signal to IIS server when stopping the event

timestampOffsetMode :: Maybe SmoothGroupTimestampOffsetMode

Type of timestamp date offset to use. - useEventStartDate: Use the date the event was started as the offset - useConfiguredOffset: Use an explicitly configured date as the offset

restartDelay :: Maybe Natural

Number of seconds before initiating a restart due to output failure, due to exhausting the numRetries on one segment, or exceeding filecacheDuration.

destination :: OutputLocationRef

Smooth Streaming publish point on an IIS server. Elemental Live acts as a "Push" encoder to IIS.

newMsSmoothGroupSettings

Arguments

:: OutputLocationRef
$sel:destination:MsSmoothGroupSettings'

-> MsSmoothGroupSettings

Create a value of MsSmoothGroupSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamManifestBehavior:MsSmoothGroupSettings', msSmoothGroupSettings_streamManifestBehavior - When set to send, send stream manifest so publishing point doesn't start until all streams start.

$sel:filecacheDuration:MsSmoothGroupSettings', msSmoothGroupSettings_filecacheDuration - Size in seconds of file cache for streaming outputs.

$sel:fragmentLength:MsSmoothGroupSettings', msSmoothGroupSettings_fragmentLength - Length of mp4 fragments to generate (in seconds). Fragment length must be compatible with GOP size and framerate.

$sel:eventId:MsSmoothGroupSettings', msSmoothGroupSettings_eventId - MS Smooth event ID to be sent to the IIS server. Should only be specified if eventIdMode is set to useConfigured.

$sel:certificateMode:MsSmoothGroupSettings', msSmoothGroupSettings_certificateMode - If set to verifyAuthenticity, verify the https certificate chain to a trusted Certificate Authority (CA). This will cause https outputs to self-signed certificates to fail.

$sel:acquisitionPointId:MsSmoothGroupSettings', msSmoothGroupSettings_acquisitionPointId - The ID to include in each message in the sparse track. Ignored if sparseTrackType is NONE.

$sel:numRetries:MsSmoothGroupSettings', msSmoothGroupSettings_numRetries - Number of retry attempts.

$sel:segmentationMode:MsSmoothGroupSettings', msSmoothGroupSettings_segmentationMode - useInputSegmentation has been deprecated. The configured segment size is always used.

$sel:audioOnlyTimecodeControl:MsSmoothGroupSettings', msSmoothGroupSettings_audioOnlyTimecodeControl - If set to passthrough for an audio-only MS Smooth output, the fragment absolute time will be set to the current timecode. This option does not write timecodes to the audio elementary stream.

$sel:eventIdMode:MsSmoothGroupSettings', msSmoothGroupSettings_eventIdMode - Specifies whether or not to send an event ID to the IIS server. If no event ID is sent and the same Live Event is used without changing the publishing point, clients might see cached video from the previous run. Options: - "useConfigured" - use the value provided in eventId - "useTimestamp" - generate and send an event ID based on the current timestamp - "noEventId" - do not send an event ID to the IIS server.

$sel:sendDelayMs:MsSmoothGroupSettings', msSmoothGroupSettings_sendDelayMs - Number of milliseconds to delay the output from the second pipeline.

$sel:sparseTrackType:MsSmoothGroupSettings', msSmoothGroupSettings_sparseTrackType - Identifies the type of data to place in the sparse track: - SCTE35: Insert SCTE-35 messages from the source content. With each message, insert an IDR frame to start a new segment. - SCTE35_WITHOUT_SEGMENTATION: Insert SCTE-35 messages from the source content. With each message, insert an IDR frame but don't start a new segment. - NONE: Don't generate a sparse track for any outputs in this output group.

$sel:connectionRetryInterval:MsSmoothGroupSettings', msSmoothGroupSettings_connectionRetryInterval - Number of seconds to wait before retrying connection to the IIS server if the connection is lost. Content will be cached during this time and the cache will be be delivered to the IIS server once the connection is re-established.

$sel:inputLossAction:MsSmoothGroupSettings', msSmoothGroupSettings_inputLossAction - Parameter that control output group behavior on input loss.

$sel:timestampOffset:MsSmoothGroupSettings', msSmoothGroupSettings_timestampOffset - Timestamp offset for the event. Only used if timestampOffsetMode is set to useConfiguredOffset.

$sel:eventStopBehavior:MsSmoothGroupSettings', msSmoothGroupSettings_eventStopBehavior - When set to sendEos, send EOS signal to IIS server when stopping the event

$sel:timestampOffsetMode:MsSmoothGroupSettings', msSmoothGroupSettings_timestampOffsetMode - Type of timestamp date offset to use. - useEventStartDate: Use the date the event was started as the offset - useConfiguredOffset: Use an explicitly configured date as the offset

$sel:restartDelay:MsSmoothGroupSettings', msSmoothGroupSettings_restartDelay - Number of seconds before initiating a restart due to output failure, due to exhausting the numRetries on one segment, or exceeding filecacheDuration.

$sel:destination:MsSmoothGroupSettings', msSmoothGroupSettings_destination - Smooth Streaming publish point on an IIS server. Elemental Live acts as a "Push" encoder to IIS.

msSmoothGroupSettings_streamManifestBehavior :: Lens' MsSmoothGroupSettings (Maybe SmoothGroupStreamManifestBehavior)

When set to send, send stream manifest so publishing point doesn't start until all streams start.

msSmoothGroupSettings_filecacheDuration :: Lens' MsSmoothGroupSettings (Maybe Natural)

Size in seconds of file cache for streaming outputs.

msSmoothGroupSettings_fragmentLength :: Lens' MsSmoothGroupSettings (Maybe Natural)

Length of mp4 fragments to generate (in seconds). Fragment length must be compatible with GOP size and framerate.

msSmoothGroupSettings_eventId :: Lens' MsSmoothGroupSettings (Maybe Text)

MS Smooth event ID to be sent to the IIS server. Should only be specified if eventIdMode is set to useConfigured.

msSmoothGroupSettings_certificateMode :: Lens' MsSmoothGroupSettings (Maybe SmoothGroupCertificateMode)

If set to verifyAuthenticity, verify the https certificate chain to a trusted Certificate Authority (CA). This will cause https outputs to self-signed certificates to fail.

msSmoothGroupSettings_acquisitionPointId :: Lens' MsSmoothGroupSettings (Maybe Text)

The ID to include in each message in the sparse track. Ignored if sparseTrackType is NONE.

msSmoothGroupSettings_numRetries :: Lens' MsSmoothGroupSettings (Maybe Natural)

Number of retry attempts.

msSmoothGroupSettings_segmentationMode :: Lens' MsSmoothGroupSettings (Maybe SmoothGroupSegmentationMode)

useInputSegmentation has been deprecated. The configured segment size is always used.

msSmoothGroupSettings_audioOnlyTimecodeControl :: Lens' MsSmoothGroupSettings (Maybe SmoothGroupAudioOnlyTimecodeControl)

If set to passthrough for an audio-only MS Smooth output, the fragment absolute time will be set to the current timecode. This option does not write timecodes to the audio elementary stream.

msSmoothGroupSettings_eventIdMode :: Lens' MsSmoothGroupSettings (Maybe SmoothGroupEventIdMode)

Specifies whether or not to send an event ID to the IIS server. If no event ID is sent and the same Live Event is used without changing the publishing point, clients might see cached video from the previous run. Options: - "useConfigured" - use the value provided in eventId - "useTimestamp" - generate and send an event ID based on the current timestamp - "noEventId" - do not send an event ID to the IIS server.

msSmoothGroupSettings_sendDelayMs :: Lens' MsSmoothGroupSettings (Maybe Natural)

Number of milliseconds to delay the output from the second pipeline.

msSmoothGroupSettings_sparseTrackType :: Lens' MsSmoothGroupSettings (Maybe SmoothGroupSparseTrackType)

Identifies the type of data to place in the sparse track: - SCTE35: Insert SCTE-35 messages from the source content. With each message, insert an IDR frame to start a new segment. - SCTE35_WITHOUT_SEGMENTATION: Insert SCTE-35 messages from the source content. With each message, insert an IDR frame but don't start a new segment. - NONE: Don't generate a sparse track for any outputs in this output group.

msSmoothGroupSettings_connectionRetryInterval :: Lens' MsSmoothGroupSettings (Maybe Natural)

Number of seconds to wait before retrying connection to the IIS server if the connection is lost. Content will be cached during this time and the cache will be be delivered to the IIS server once the connection is re-established.

msSmoothGroupSettings_inputLossAction :: Lens' MsSmoothGroupSettings (Maybe InputLossActionForMsSmoothOut)

Parameter that control output group behavior on input loss.

msSmoothGroupSettings_timestampOffset :: Lens' MsSmoothGroupSettings (Maybe Text)

Timestamp offset for the event. Only used if timestampOffsetMode is set to useConfiguredOffset.

msSmoothGroupSettings_eventStopBehavior :: Lens' MsSmoothGroupSettings (Maybe SmoothGroupEventStopBehavior)

When set to sendEos, send EOS signal to IIS server when stopping the event

msSmoothGroupSettings_timestampOffsetMode :: Lens' MsSmoothGroupSettings (Maybe SmoothGroupTimestampOffsetMode)

Type of timestamp date offset to use. - useEventStartDate: Use the date the event was started as the offset - useConfiguredOffset: Use an explicitly configured date as the offset

msSmoothGroupSettings_restartDelay :: Lens' MsSmoothGroupSettings (Maybe Natural)

Number of seconds before initiating a restart due to output failure, due to exhausting the numRetries on one segment, or exceeding filecacheDuration.

msSmoothGroupSettings_destination :: Lens' MsSmoothGroupSettings OutputLocationRef

Smooth Streaming publish point on an IIS server. Elemental Live acts as a "Push" encoder to IIS.