Network.AWS.MediaLive.Types.Multiplex

Description

Documentation

data Multiplex

The multiplex object.

See: newMultiplex smart constructor.

Constructors

Multiplex'

Fields

state :: Maybe MultiplexState

The current state of the multiplex.

arn :: Maybe Text

The unique arn of the multiplex.

pipelinesRunningCount :: Maybe Int

The number of currently healthy pipelines.

availabilityZones :: Maybe [Text]

A list of availability zones for the multiplex.

programCount :: Maybe Int

The number of programs in the multiplex.

destinations :: Maybe [MultiplexOutputDestination]

A list of the multiplex output destinations.

name :: Maybe Text

The name of the multiplex.

id :: Maybe Text

The unique id of the multiplex.

multiplexSettings :: Maybe MultiplexSettings

Configuration for a multiplex event.

tags :: Maybe (HashMap Text Text)

A collection of key-value pairs.

newMultiplex :: Multiplex

Create a value of Multiplex with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:Multiplex', multiplex_state - The current state of the multiplex.

$sel:arn:Multiplex', multiplex_arn - The unique arn of the multiplex.

$sel:pipelinesRunningCount:Multiplex', multiplex_pipelinesRunningCount - The number of currently healthy pipelines.

$sel:availabilityZones:Multiplex', multiplex_availabilityZones - A list of availability zones for the multiplex.

$sel:programCount:Multiplex', multiplex_programCount - The number of programs in the multiplex.

$sel:destinations:Multiplex', multiplex_destinations - A list of the multiplex output destinations.

$sel:name:Multiplex', multiplex_name - The name of the multiplex.

$sel:id:Multiplex', multiplex_id - The unique id of the multiplex.

$sel:multiplexSettings:Multiplex', multiplex_multiplexSettings - Configuration for a multiplex event.

$sel:tags:Multiplex', multiplex_tags - A collection of key-value pairs.

multiplex_state :: Lens' Multiplex (Maybe MultiplexState)

The current state of the multiplex.

multiplex_arn :: Lens' Multiplex (Maybe Text)

The unique arn of the multiplex.

multiplex_pipelinesRunningCount :: Lens' Multiplex (Maybe Int)

The number of currently healthy pipelines.

multiplex_availabilityZones :: Lens' Multiplex (Maybe [Text])

A list of availability zones for the multiplex.

multiplex_programCount :: Lens' Multiplex (Maybe Int)

The number of programs in the multiplex.

multiplex_destinations :: Lens' Multiplex (Maybe [MultiplexOutputDestination])

A list of the multiplex output destinations.

multiplex_name :: Lens' Multiplex (Maybe Text)

The name of the multiplex.

multiplex_id :: Lens' Multiplex (Maybe Text)

The unique id of the multiplex.

multiplex_multiplexSettings :: Lens' Multiplex (Maybe MultiplexSettings)

Configuration for a multiplex event.

multiplex_tags :: Lens' Multiplex (Maybe (HashMap Text Text))

A collection of key-value pairs.