Network.AWS.MediaLive.Types.Offering

Description

Documentation

data Offering

Reserved resources available for purchase

See: newOffering smart constructor.

Constructors

Offering'

Fields

resourceSpecification :: Maybe ReservationResourceSpecification

Resource configuration details

currencyCode :: Maybe Text

Currency code for usagePrice and fixedPrice in ISO-4217 format, e.g. 'USD'

arn :: Maybe Text

Unique offering ARN, e.g. 'arn:aws:medialive:us-west-2:123456789012:offering:87654321'

offeringId :: Maybe Text

Unique offering ID, e.g. '87654321'

region :: Maybe Text

AWS region, e.g. 'us-west-2'

offeringType :: Maybe OfferingType

Offering type, e.g. 'NO_UPFRONT'

usagePrice :: Maybe Double

Recurring usage charge for each reserved resource, e.g. '157.0'

fixedPrice :: Maybe Double

One-time charge for each reserved resource, e.g. '0.0' for a NO_UPFRONT offering

durationUnits :: Maybe OfferingDurationUnits

Units for duration, e.g. 'MONTHS'

offeringDescription :: Maybe Text

Offering description, e.g. 'HD AVC output at 10-20 Mbps, 30 fps, and standard VQ in US West (Oregon)'

duration :: Maybe Int

Lease duration, e.g. '12'

newOffering :: Offering

Create a value of Offering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceSpecification:Offering', offering_resourceSpecification - Resource configuration details

$sel:currencyCode:Offering', offering_currencyCode - Currency code for usagePrice and fixedPrice in ISO-4217 format, e.g. 'USD'

$sel:arn:Offering', offering_arn - Unique offering ARN, e.g. 'arn:aws:medialive:us-west-2:123456789012:offering:87654321'

$sel:offeringId:Offering', offering_offeringId - Unique offering ID, e.g. '87654321'

$sel:region:Offering', offering_region - AWS region, e.g. 'us-west-2'

$sel:offeringType:Offering', offering_offeringType - Offering type, e.g. 'NO_UPFRONT'

$sel:usagePrice:Offering', offering_usagePrice - Recurring usage charge for each reserved resource, e.g. '157.0'

$sel:fixedPrice:Offering', offering_fixedPrice - One-time charge for each reserved resource, e.g. '0.0' for a NO_UPFRONT offering

$sel:durationUnits:Offering', offering_durationUnits - Units for duration, e.g. 'MONTHS'

$sel:offeringDescription:Offering', offering_offeringDescription - Offering description, e.g. 'HD AVC output at 10-20 Mbps, 30 fps, and standard VQ in US West (Oregon)'

$sel:duration:Offering', offering_duration - Lease duration, e.g. '12'

offering_resourceSpecification :: Lens' Offering (Maybe ReservationResourceSpecification)

Resource configuration details

offering_currencyCode :: Lens' Offering (Maybe Text)

Currency code for usagePrice and fixedPrice in ISO-4217 format, e.g. 'USD'

offering_arn :: Lens' Offering (Maybe Text)

Unique offering ARN, e.g. 'arn:aws:medialive:us-west-2:123456789012:offering:87654321'

offering_offeringId :: Lens' Offering (Maybe Text)

Unique offering ID, e.g. '87654321'

offering_region :: Lens' Offering (Maybe Text)

AWS region, e.g. 'us-west-2'

offering_offeringType :: Lens' Offering (Maybe OfferingType)

Offering type, e.g. 'NO_UPFRONT'

offering_usagePrice :: Lens' Offering (Maybe Double)

Recurring usage charge for each reserved resource, e.g. '157.0'

offering_fixedPrice :: Lens' Offering (Maybe Double)

One-time charge for each reserved resource, e.g. '0.0' for a NO_UPFRONT offering

offering_durationUnits :: Lens' Offering (Maybe OfferingDurationUnits)

Units for duration, e.g. 'MONTHS'

offering_offeringDescription :: Lens' Offering (Maybe Text)

Offering description, e.g. 'HD AVC output at 10-20 Mbps, 30 fps, and standard VQ in US West (Oregon)'

offering_duration :: Lens' Offering (Maybe Int)

Lease duration, e.g. '12'