Network.AWS.MediaLive.Types.Reservation

Description

Documentation

data Reservation

Reserved resources available to use

See: newReservation smart constructor.

Constructors

Reservation'

Fields

state :: Maybe ReservationState

Current state of reservation, e.g. 'ACTIVE'

resourceSpecification :: Maybe ReservationResourceSpecification

Resource configuration details

currencyCode :: Maybe Text

Currency code for usagePrice and fixedPrice in ISO-4217 format, e.g. 'USD'

arn :: Maybe Text

Unique reservation ARN, e.g. 'arn:aws:medialive:us-west-2:123456789012:reservation:1234567'

start :: Maybe Text

Reservation UTC start date and time in ISO-8601 format, e.g. '2018-03-01T00:00:00'

count :: Maybe Int

Number of reserved resources

end :: Maybe Text

Reservation UTC end date and time in ISO-8601 format, e.g. '2019-03-01T00:00:00'

name :: Maybe Text

User specified reservation name

reservationId :: Maybe Text

Unique reservation ID, e.g. '1234567'

offeringId :: Maybe Text

Unique offering ID, e.g. '87654321'

region :: Maybe Text

AWS region, e.g. 'us-west-2'

offeringType :: Maybe OfferingType

Offering type, e.g. 'NO_UPFRONT'

usagePrice :: Maybe Double

Recurring usage charge for each reserved resource, e.g. '157.0'

fixedPrice :: Maybe Double

One-time charge for each reserved resource, e.g. '0.0' for a NO_UPFRONT offering

durationUnits :: Maybe OfferingDurationUnits

Units for duration, e.g. 'MONTHS'

offeringDescription :: Maybe Text

Offering description, e.g. 'HD AVC output at 10-20 Mbps, 30 fps, and standard VQ in US West (Oregon)'

duration :: Maybe Int

Lease duration, e.g. '12'

tags :: Maybe (HashMap Text Text)

A collection of key-value pairs

newReservation :: Reservation

Create a value of Reservation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:Reservation', reservation_state - Current state of reservation, e.g. 'ACTIVE'

$sel:resourceSpecification:Reservation', reservation_resourceSpecification - Resource configuration details

$sel:currencyCode:Reservation', reservation_currencyCode - Currency code for usagePrice and fixedPrice in ISO-4217 format, e.g. 'USD'

$sel:arn:Reservation', reservation_arn - Unique reservation ARN, e.g. 'arn:aws:medialive:us-west-2:123456789012:reservation:1234567'

$sel:start:Reservation', reservation_start - Reservation UTC start date and time in ISO-8601 format, e.g. '2018-03-01T00:00:00'

$sel:count:Reservation', reservation_count - Number of reserved resources

$sel:end:Reservation', reservation_end - Reservation UTC end date and time in ISO-8601 format, e.g. '2019-03-01T00:00:00'

$sel:name:Reservation', reservation_name - User specified reservation name

$sel:reservationId:Reservation', reservation_reservationId - Unique reservation ID, e.g. '1234567'

$sel:offeringId:Reservation', reservation_offeringId - Unique offering ID, e.g. '87654321'

$sel:region:Reservation', reservation_region - AWS region, e.g. 'us-west-2'

$sel:offeringType:Reservation', reservation_offeringType - Offering type, e.g. 'NO_UPFRONT'

$sel:usagePrice:Reservation', reservation_usagePrice - Recurring usage charge for each reserved resource, e.g. '157.0'

$sel:fixedPrice:Reservation', reservation_fixedPrice - One-time charge for each reserved resource, e.g. '0.0' for a NO_UPFRONT offering

$sel:durationUnits:Reservation', reservation_durationUnits - Units for duration, e.g. 'MONTHS'

$sel:offeringDescription:Reservation', reservation_offeringDescription - Offering description, e.g. 'HD AVC output at 10-20 Mbps, 30 fps, and standard VQ in US West (Oregon)'

$sel:duration:Reservation', reservation_duration - Lease duration, e.g. '12'

$sel:tags:Reservation', reservation_tags - A collection of key-value pairs

reservation_state :: Lens' Reservation (Maybe ReservationState)

Current state of reservation, e.g. 'ACTIVE'

reservation_resourceSpecification :: Lens' Reservation (Maybe ReservationResourceSpecification)

Resource configuration details

reservation_currencyCode :: Lens' Reservation (Maybe Text)

Currency code for usagePrice and fixedPrice in ISO-4217 format, e.g. 'USD'

reservation_arn :: Lens' Reservation (Maybe Text)

Unique reservation ARN, e.g. 'arn:aws:medialive:us-west-2:123456789012:reservation:1234567'

reservation_start :: Lens' Reservation (Maybe Text)

Reservation UTC start date and time in ISO-8601 format, e.g. '2018-03-01T00:00:00'

reservation_count :: Lens' Reservation (Maybe Int)

Number of reserved resources

reservation_end :: Lens' Reservation (Maybe Text)

Reservation UTC end date and time in ISO-8601 format, e.g. '2019-03-01T00:00:00'

reservation_name :: Lens' Reservation (Maybe Text)

User specified reservation name

reservation_reservationId :: Lens' Reservation (Maybe Text)

Unique reservation ID, e.g. '1234567'

reservation_offeringId :: Lens' Reservation (Maybe Text)

Unique offering ID, e.g. '87654321'

reservation_region :: Lens' Reservation (Maybe Text)

AWS region, e.g. 'us-west-2'

reservation_offeringType :: Lens' Reservation (Maybe OfferingType)

Offering type, e.g. 'NO_UPFRONT'

reservation_usagePrice :: Lens' Reservation (Maybe Double)

Recurring usage charge for each reserved resource, e.g. '157.0'

reservation_fixedPrice :: Lens' Reservation (Maybe Double)

One-time charge for each reserved resource, e.g. '0.0' for a NO_UPFRONT offering

reservation_durationUnits :: Lens' Reservation (Maybe OfferingDurationUnits)

Units for duration, e.g. 'MONTHS'

reservation_offeringDescription :: Lens' Reservation (Maybe Text)

Offering description, e.g. 'HD AVC output at 10-20 Mbps, 30 fps, and standard VQ in US West (Oregon)'

reservation_duration :: Lens' Reservation (Maybe Int)

Lease duration, e.g. '12'

reservation_tags :: Lens' Reservation (Maybe (HashMap Text Text))

A collection of key-value pairs