Network.AWS.MediaLive.Types.RtmpGroupSettings

Description

Documentation

data RtmpGroupSettings

Rtmp Group Settings

See: newRtmpGroupSettings smart constructor.

Constructors

RtmpGroupSettings'

Fields

adMarkers :: Maybe [RtmpAdMarkers]

Choose the ad marker type for this output group. MediaLive will create a message based on the content of each SCTE-35 message, format it for that marker type, and insert it in the datastream.

captionData :: Maybe RtmpCaptionData

Controls the types of data that passes to onCaptionInfo outputs. If set to 'all' then 608 and 708 carried DTVCC data will be passed. If set to 'field1AndField2608' then DTVCC data will be stripped out, but 608 data from both fields will be passed. If set to 'field1608' then only the data carried in 608 from field 1 video will be passed.

cacheFullBehavior :: Maybe RtmpCacheFullBehavior

Controls behavior when content cache fills up. If remote origin server stalls the RTMP connection and does not accept content fast enough the 'Media Cache' will fill up. When the cache reaches the duration specified by cacheLength the cache will stop accepting new content. If set to disconnectImmediately, the RTMP output will force a disconnect. Clear the media cache, and reconnect after restartDelay seconds. If set to waitForServer, the RTMP output will wait up to 5 minutes to allow the origin server to begin accepting data again.

cacheLength :: Maybe Natural

Cache length, in seconds, is used to calculate buffer size.

authenticationScheme :: Maybe AuthenticationScheme

Authentication scheme to use when connecting with CDN

inputLossAction :: Maybe InputLossActionForRtmpOut

Controls the behavior of this RTMP group if input becomes unavailable. - emitOutput: Emit a slate until input returns. - pauseOutput: Stop transmitting data until input returns. This does not close the underlying RTMP connection.

restartDelay :: Maybe Natural

If a streaming output fails, number of seconds to wait until a restart is initiated. A value of 0 means never restart.

newRtmpGroupSettings :: RtmpGroupSettings

Create a value of RtmpGroupSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:adMarkers:RtmpGroupSettings', rtmpGroupSettings_adMarkers - Choose the ad marker type for this output group. MediaLive will create a message based on the content of each SCTE-35 message, format it for that marker type, and insert it in the datastream.

$sel:captionData:RtmpGroupSettings', rtmpGroupSettings_captionData - Controls the types of data that passes to onCaptionInfo outputs. If set to 'all' then 608 and 708 carried DTVCC data will be passed. If set to 'field1AndField2608' then DTVCC data will be stripped out, but 608 data from both fields will be passed. If set to 'field1608' then only the data carried in 608 from field 1 video will be passed.

$sel:cacheFullBehavior:RtmpGroupSettings', rtmpGroupSettings_cacheFullBehavior - Controls behavior when content cache fills up. If remote origin server stalls the RTMP connection and does not accept content fast enough the 'Media Cache' will fill up. When the cache reaches the duration specified by cacheLength the cache will stop accepting new content. If set to disconnectImmediately, the RTMP output will force a disconnect. Clear the media cache, and reconnect after restartDelay seconds. If set to waitForServer, the RTMP output will wait up to 5 minutes to allow the origin server to begin accepting data again.

$sel:cacheLength:RtmpGroupSettings', rtmpGroupSettings_cacheLength - Cache length, in seconds, is used to calculate buffer size.

$sel:authenticationScheme:RtmpGroupSettings', rtmpGroupSettings_authenticationScheme - Authentication scheme to use when connecting with CDN

$sel:inputLossAction:RtmpGroupSettings', rtmpGroupSettings_inputLossAction - Controls the behavior of this RTMP group if input becomes unavailable. - emitOutput: Emit a slate until input returns. - pauseOutput: Stop transmitting data until input returns. This does not close the underlying RTMP connection.

$sel:restartDelay:RtmpGroupSettings', rtmpGroupSettings_restartDelay - If a streaming output fails, number of seconds to wait until a restart is initiated. A value of 0 means never restart.

rtmpGroupSettings_adMarkers :: Lens' RtmpGroupSettings (Maybe [RtmpAdMarkers])

Choose the ad marker type for this output group. MediaLive will create a message based on the content of each SCTE-35 message, format it for that marker type, and insert it in the datastream.

rtmpGroupSettings_captionData :: Lens' RtmpGroupSettings (Maybe RtmpCaptionData)

Controls the types of data that passes to onCaptionInfo outputs. If set to 'all' then 608 and 708 carried DTVCC data will be passed. If set to 'field1AndField2608' then DTVCC data will be stripped out, but 608 data from both fields will be passed. If set to 'field1608' then only the data carried in 608 from field 1 video will be passed.

rtmpGroupSettings_cacheFullBehavior :: Lens' RtmpGroupSettings (Maybe RtmpCacheFullBehavior)

Controls behavior when content cache fills up. If remote origin server stalls the RTMP connection and does not accept content fast enough the 'Media Cache' will fill up. When the cache reaches the duration specified by cacheLength the cache will stop accepting new content. If set to disconnectImmediately, the RTMP output will force a disconnect. Clear the media cache, and reconnect after restartDelay seconds. If set to waitForServer, the RTMP output will wait up to 5 minutes to allow the origin server to begin accepting data again.

rtmpGroupSettings_cacheLength :: Lens' RtmpGroupSettings (Maybe Natural)

Cache length, in seconds, is used to calculate buffer size.

rtmpGroupSettings_authenticationScheme :: Lens' RtmpGroupSettings (Maybe AuthenticationScheme)

Authentication scheme to use when connecting with CDN

rtmpGroupSettings_inputLossAction :: Lens' RtmpGroupSettings (Maybe InputLossActionForRtmpOut)

Controls the behavior of this RTMP group if input becomes unavailable. - emitOutput: Emit a slate until input returns. - pauseOutput: Stop transmitting data until input returns. This does not close the underlying RTMP connection.

rtmpGroupSettings_restartDelay :: Lens' RtmpGroupSettings (Maybe Natural)

If a streaming output fails, number of seconds to wait until a restart is initiated. A value of 0 means never restart.