Network.AWS.MediaLive.Types.ScheduleActionSettings

Description

Documentation

data ScheduleActionSettings

Holds the settings for a single schedule action.

See: newScheduleActionSettings smart constructor.

Constructors

ScheduleActionSettings'

Fields

staticImageDeactivateSettings :: Maybe StaticImageDeactivateScheduleActionSettings

Action to deactivate a static image overlay

scte35SpliceInsertSettings :: Maybe Scte35SpliceInsertScheduleActionSettings

Action to insert SCTE-35 splice_insert message

motionGraphicsImageActivateSettings :: Maybe MotionGraphicsActivateScheduleActionSettings

Action to activate a motion graphics image overlay

staticImageActivateSettings :: Maybe StaticImageActivateScheduleActionSettings

Action to activate a static image overlay

motionGraphicsImageDeactivateSettings :: Maybe MotionGraphicsDeactivateScheduleActionSettings

Action to deactivate a motion graphics image overlay

scte35TimeSignalSettings :: Maybe Scte35TimeSignalScheduleActionSettings

Action to insert SCTE-35 time_signal message

inputPrepareSettings :: Maybe InputPrepareScheduleActionSettings

Action to prepare an input for a future immediate input switch

hlsId3SegmentTaggingSettings :: Maybe HlsId3SegmentTaggingScheduleActionSettings

Action to insert HLS ID3 segment tagging

scte35ReturnToNetworkSettings :: Maybe Scte35ReturnToNetworkScheduleActionSettings

Action to insert SCTE-35 return_to_network message

pauseStateSettings :: Maybe PauseStateScheduleActionSettings

Action to pause or unpause one or both channel pipelines

hlsTimedMetadataSettings :: Maybe HlsTimedMetadataScheduleActionSettings

Action to insert HLS metadata

inputSwitchSettings :: Maybe InputSwitchScheduleActionSettings

Action to switch the input

newScheduleActionSettings :: ScheduleActionSettings

Create a value of ScheduleActionSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:staticImageDeactivateSettings:ScheduleActionSettings', scheduleActionSettings_staticImageDeactivateSettings - Action to deactivate a static image overlay

$sel:scte35SpliceInsertSettings:ScheduleActionSettings', scheduleActionSettings_scte35SpliceInsertSettings - Action to insert SCTE-35 splice_insert message

$sel:motionGraphicsImageActivateSettings:ScheduleActionSettings', scheduleActionSettings_motionGraphicsImageActivateSettings - Action to activate a motion graphics image overlay

$sel:staticImageActivateSettings:ScheduleActionSettings', scheduleActionSettings_staticImageActivateSettings - Action to activate a static image overlay

$sel:motionGraphicsImageDeactivateSettings:ScheduleActionSettings', scheduleActionSettings_motionGraphicsImageDeactivateSettings - Action to deactivate a motion graphics image overlay

$sel:scte35TimeSignalSettings:ScheduleActionSettings', scheduleActionSettings_scte35TimeSignalSettings - Action to insert SCTE-35 time_signal message

$sel:inputPrepareSettings:ScheduleActionSettings', scheduleActionSettings_inputPrepareSettings - Action to prepare an input for a future immediate input switch

$sel:hlsId3SegmentTaggingSettings:ScheduleActionSettings', scheduleActionSettings_hlsId3SegmentTaggingSettings - Action to insert HLS ID3 segment tagging

$sel:scte35ReturnToNetworkSettings:ScheduleActionSettings', scheduleActionSettings_scte35ReturnToNetworkSettings - Action to insert SCTE-35 return_to_network message

$sel:pauseStateSettings:ScheduleActionSettings', scheduleActionSettings_pauseStateSettings - Action to pause or unpause one or both channel pipelines

$sel:hlsTimedMetadataSettings:ScheduleActionSettings', scheduleActionSettings_hlsTimedMetadataSettings - Action to insert HLS metadata

$sel:inputSwitchSettings:ScheduleActionSettings', scheduleActionSettings_inputSwitchSettings - Action to switch the input

scheduleActionSettings_staticImageDeactivateSettings :: Lens' ScheduleActionSettings (Maybe StaticImageDeactivateScheduleActionSettings)

Action to deactivate a static image overlay

scheduleActionSettings_scte35SpliceInsertSettings :: Lens' ScheduleActionSettings (Maybe Scte35SpliceInsertScheduleActionSettings)

Action to insert SCTE-35 splice_insert message

scheduleActionSettings_motionGraphicsImageActivateSettings :: Lens' ScheduleActionSettings (Maybe MotionGraphicsActivateScheduleActionSettings)

Action to activate a motion graphics image overlay

scheduleActionSettings_staticImageActivateSettings :: Lens' ScheduleActionSettings (Maybe StaticImageActivateScheduleActionSettings)

Action to activate a static image overlay

scheduleActionSettings_motionGraphicsImageDeactivateSettings :: Lens' ScheduleActionSettings (Maybe MotionGraphicsDeactivateScheduleActionSettings)

Action to deactivate a motion graphics image overlay

scheduleActionSettings_scte35TimeSignalSettings :: Lens' ScheduleActionSettings (Maybe Scte35TimeSignalScheduleActionSettings)

Action to insert SCTE-35 time_signal message

scheduleActionSettings_inputPrepareSettings :: Lens' ScheduleActionSettings (Maybe InputPrepareScheduleActionSettings)

Action to prepare an input for a future immediate input switch

scheduleActionSettings_hlsId3SegmentTaggingSettings :: Lens' ScheduleActionSettings (Maybe HlsId3SegmentTaggingScheduleActionSettings)

Action to insert HLS ID3 segment tagging

scheduleActionSettings_scte35ReturnToNetworkSettings :: Lens' ScheduleActionSettings (Maybe Scte35ReturnToNetworkScheduleActionSettings)

Action to insert SCTE-35 return_to_network message

scheduleActionSettings_pauseStateSettings :: Lens' ScheduleActionSettings (Maybe PauseStateScheduleActionSettings)

Action to pause or unpause one or both channel pipelines

scheduleActionSettings_hlsTimedMetadataSettings :: Lens' ScheduleActionSettings (Maybe HlsTimedMetadataScheduleActionSettings)

Action to insert HLS metadata

scheduleActionSettings_inputSwitchSettings :: Lens' ScheduleActionSettings (Maybe InputSwitchScheduleActionSettings)

Action to switch the input