Network.AWS.MediaLive.Types.Scte27SourceSettings

Description

Documentation

data Scte27SourceSettings

Scte27 Source Settings

See: newScte27SourceSettings smart constructor.

Constructors

Scte27SourceSettings'

Fields

ocrLanguage :: Maybe Scte27OcrLanguage

If you will configure a WebVTT caption description that references this caption selector, use this field to provide the language to consider when translating the image-based source to text.

pid :: Maybe Natural

The pid field is used in conjunction with the caption selector languageCode field as follows: - Specify PID and Language: Extracts captions from that PID; the language is "informational". - Specify PID and omit Language: Extracts the specified PID. - Omit PID and specify Language: Extracts the specified language, whichever PID that happens to be. - Omit PID and omit Language: Valid only if source is DVB-Sub that is being passed through; all languages will be passed through.

newScte27SourceSettings :: Scte27SourceSettings

Create a value of Scte27SourceSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ocrLanguage:Scte27SourceSettings', scte27SourceSettings_ocrLanguage - If you will configure a WebVTT caption description that references this caption selector, use this field to provide the language to consider when translating the image-based source to text.

$sel:pid:Scte27SourceSettings', scte27SourceSettings_pid - The pid field is used in conjunction with the caption selector languageCode field as follows: - Specify PID and Language: Extracts captions from that PID; the language is "informational". - Specify PID and omit Language: Extracts the specified PID. - Omit PID and specify Language: Extracts the specified language, whichever PID that happens to be. - Omit PID and omit Language: Valid only if source is DVB-Sub that is being passed through; all languages will be passed through.

scte27SourceSettings_ocrLanguage :: Lens' Scte27SourceSettings (Maybe Scte27OcrLanguage)

If you will configure a WebVTT caption description that references this caption selector, use this field to provide the language to consider when translating the image-based source to text.

scte27SourceSettings_pid :: Lens' Scte27SourceSettings (Maybe Natural)

The pid field is used in conjunction with the caption selector languageCode field as follows: - Specify PID and Language: Extracts captions from that PID; the language is "informational". - Specify PID and omit Language: Extracts the specified PID. - Omit PID and specify Language: Extracts the specified language, whichever PID that happens to be. - Omit PID and omit Language: Valid only if source is DVB-Sub that is being passed through; all languages will be passed through.