Network.AWS.MediaLive.Types.Scte35DeliveryRestrictions

Description

Documentation

data Scte35DeliveryRestrictions

Corresponds to SCTE-35 delivery_not_restricted_flag parameter. To declare delivery restrictions, include this element and its four "restriction" flags. To declare that there are no restrictions, omit this element.

See: newScte35DeliveryRestrictions smart constructor.

Constructors

Scte35DeliveryRestrictions'

Fields

deviceRestrictions :: Scte35DeviceRestrictions

Corresponds to SCTE-35 device_restrictions parameter.

archiveAllowedFlag :: Scte35ArchiveAllowedFlag

Corresponds to SCTE-35 archive_allowed_flag.

webDeliveryAllowedFlag :: Scte35WebDeliveryAllowedFlag

Corresponds to SCTE-35 web_delivery_allowed_flag parameter.

noRegionalBlackoutFlag :: Scte35NoRegionalBlackoutFlag

Corresponds to SCTE-35 no_regional_blackout_flag parameter.

newScte35DeliveryRestrictions

Arguments

:: Scte35DeviceRestrictions
$sel:deviceRestrictions:Scte35DeliveryRestrictions'

-> Scte35ArchiveAllowedFlag
$sel:archiveAllowedFlag:Scte35DeliveryRestrictions'

-> Scte35WebDeliveryAllowedFlag
$sel:webDeliveryAllowedFlag:Scte35DeliveryRestrictions'

-> Scte35NoRegionalBlackoutFlag
$sel:noRegionalBlackoutFlag:Scte35DeliveryRestrictions'

-> Scte35DeliveryRestrictions

Create a value of Scte35DeliveryRestrictions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deviceRestrictions:Scte35DeliveryRestrictions', scte35DeliveryRestrictions_deviceRestrictions - Corresponds to SCTE-35 device_restrictions parameter.

$sel:archiveAllowedFlag:Scte35DeliveryRestrictions', scte35DeliveryRestrictions_archiveAllowedFlag - Corresponds to SCTE-35 archive_allowed_flag.

$sel:webDeliveryAllowedFlag:Scte35DeliveryRestrictions', scte35DeliveryRestrictions_webDeliveryAllowedFlag - Corresponds to SCTE-35 web_delivery_allowed_flag parameter.

$sel:noRegionalBlackoutFlag:Scte35DeliveryRestrictions', scte35DeliveryRestrictions_noRegionalBlackoutFlag - Corresponds to SCTE-35 no_regional_blackout_flag parameter.

scte35DeliveryRestrictions_deviceRestrictions :: Lens' Scte35DeliveryRestrictions Scte35DeviceRestrictions

Corresponds to SCTE-35 device_restrictions parameter.

scte35DeliveryRestrictions_archiveAllowedFlag :: Lens' Scte35DeliveryRestrictions Scte35ArchiveAllowedFlag

Corresponds to SCTE-35 archive_allowed_flag.

scte35DeliveryRestrictions_webDeliveryAllowedFlag :: Lens' Scte35DeliveryRestrictions Scte35WebDeliveryAllowedFlag

Corresponds to SCTE-35 web_delivery_allowed_flag parameter.

scte35DeliveryRestrictions_noRegionalBlackoutFlag :: Lens' Scte35DeliveryRestrictions Scte35NoRegionalBlackoutFlag

Corresponds to SCTE-35 no_regional_blackout_flag parameter.