Network.AWS.MediaLive.Types.Scte35SegmentationDescriptor

Description

Documentation

data Scte35SegmentationDescriptor

Corresponds to SCTE-35 segmentation_descriptor.

See: newScte35SegmentationDescriptor smart constructor.

Constructors

Scte35SegmentationDescriptor'

Fields

subSegmentNum :: Maybe Natural

Corresponds to SCTE-35 sub_segment_num. A value that is valid for the specified segmentation_type_id.

segmentationUpid :: Maybe Text

Corresponds to SCTE-35 segmentation_upid. Enter a string containing the hexadecimal representation of the characters that make up the SCTE-35 segmentation_upid value. Must contain an even number of hex characters. Do not include spaces between each hex pair. For example, the ASCII "ADS Information" becomes hex "41445320496e666f726d6174696f6e.

segmentNum :: Maybe Natural

Corresponds to SCTE-35 segment_num. A value that is valid for the specified segmentation_type_id.

subSegmentsExpected :: Maybe Natural

Corresponds to SCTE-35 sub_segments_expected. A value that is valid for the specified segmentation_type_id.

segmentationUpidType :: Maybe Natural

Corresponds to SCTE-35 segmentation_upid_type. On the console, enter one of the types listed in the SCTE-35 specification, converted to a decimal. For example, "0x0C" hex from the specification is "12" in decimal. In the CLI, API, or an SDK, enter one of the types listed in the SCTE-35 specification, in either hex (for example, "0x0C" ) or in decimal (for example, "12").

segmentsExpected :: Maybe Natural

Corresponds to SCTE-35 segments_expected. A value that is valid for the specified segmentation_type_id.

segmentationTypeId :: Maybe Natural

Corresponds to SCTE-35 segmentation_type_id. One of the segmentation_type_id values listed in the SCTE-35 specification. On the console, enter the ID in decimal (for example, "52"). In the CLI, API, or an SDK, enter the ID in hex (for example, "0x34") or decimal (for example, "52").

deliveryRestrictions :: Maybe Scte35DeliveryRestrictions

Holds the four SCTE-35 delivery restriction parameters.

segmentationDuration :: Maybe Natural

Corresponds to SCTE-35 segmentation_duration. Optional. The duration for the time_signal, in 90 KHz ticks. To convert seconds to ticks, multiple the seconds by 90,000. Enter time in 90 KHz clock ticks. If you do not enter a duration, the time_signal will continue until you insert a cancellation message.

segmentationEventId :: Natural

Corresponds to SCTE-35 segmentation_event_id.

segmentationCancelIndicator :: Scte35SegmentationCancelIndicator

Corresponds to SCTE-35 segmentation_event_cancel_indicator.

newScte35SegmentationDescriptor

Arguments

:: Natural
$sel:segmentationEventId:Scte35SegmentationDescriptor'

-> Scte35SegmentationCancelIndicator
$sel:segmentationCancelIndicator:Scte35SegmentationDescriptor'

-> Scte35SegmentationDescriptor

Create a value of Scte35SegmentationDescriptor with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subSegmentNum:Scte35SegmentationDescriptor', scte35SegmentationDescriptor_subSegmentNum - Corresponds to SCTE-35 sub_segment_num. A value that is valid for the specified segmentation_type_id.

$sel:segmentationUpid:Scte35SegmentationDescriptor', scte35SegmentationDescriptor_segmentationUpid - Corresponds to SCTE-35 segmentation_upid. Enter a string containing the hexadecimal representation of the characters that make up the SCTE-35 segmentation_upid value. Must contain an even number of hex characters. Do not include spaces between each hex pair. For example, the ASCII "ADS Information" becomes hex "41445320496e666f726d6174696f6e.

$sel:segmentNum:Scte35SegmentationDescriptor', scte35SegmentationDescriptor_segmentNum - Corresponds to SCTE-35 segment_num. A value that is valid for the specified segmentation_type_id.

$sel:subSegmentsExpected:Scte35SegmentationDescriptor', scte35SegmentationDescriptor_subSegmentsExpected - Corresponds to SCTE-35 sub_segments_expected. A value that is valid for the specified segmentation_type_id.

$sel:segmentationUpidType:Scte35SegmentationDescriptor', scte35SegmentationDescriptor_segmentationUpidType - Corresponds to SCTE-35 segmentation_upid_type. On the console, enter one of the types listed in the SCTE-35 specification, converted to a decimal. For example, "0x0C" hex from the specification is "12" in decimal. In the CLI, API, or an SDK, enter one of the types listed in the SCTE-35 specification, in either hex (for example, "0x0C" ) or in decimal (for example, "12").

$sel:segmentsExpected:Scte35SegmentationDescriptor', scte35SegmentationDescriptor_segmentsExpected - Corresponds to SCTE-35 segments_expected. A value that is valid for the specified segmentation_type_id.

$sel:segmentationTypeId:Scte35SegmentationDescriptor', scte35SegmentationDescriptor_segmentationTypeId - Corresponds to SCTE-35 segmentation_type_id. One of the segmentation_type_id values listed in the SCTE-35 specification. On the console, enter the ID in decimal (for example, "52"). In the CLI, API, or an SDK, enter the ID in hex (for example, "0x34") or decimal (for example, "52").

$sel:deliveryRestrictions:Scte35SegmentationDescriptor', scte35SegmentationDescriptor_deliveryRestrictions - Holds the four SCTE-35 delivery restriction parameters.

$sel:segmentationDuration:Scte35SegmentationDescriptor', scte35SegmentationDescriptor_segmentationDuration - Corresponds to SCTE-35 segmentation_duration. Optional. The duration for the time_signal, in 90 KHz ticks. To convert seconds to ticks, multiple the seconds by 90,000. Enter time in 90 KHz clock ticks. If you do not enter a duration, the time_signal will continue until you insert a cancellation message.

$sel:segmentationEventId:Scte35SegmentationDescriptor', scte35SegmentationDescriptor_segmentationEventId - Corresponds to SCTE-35 segmentation_event_id.

$sel:segmentationCancelIndicator:Scte35SegmentationDescriptor', scte35SegmentationDescriptor_segmentationCancelIndicator - Corresponds to SCTE-35 segmentation_event_cancel_indicator.

scte35SegmentationDescriptor_subSegmentNum :: Lens' Scte35SegmentationDescriptor (Maybe Natural)

Corresponds to SCTE-35 sub_segment_num. A value that is valid for the specified segmentation_type_id.

scte35SegmentationDescriptor_segmentationUpid :: Lens' Scte35SegmentationDescriptor (Maybe Text)

Corresponds to SCTE-35 segmentation_upid. Enter a string containing the hexadecimal representation of the characters that make up the SCTE-35 segmentation_upid value. Must contain an even number of hex characters. Do not include spaces between each hex pair. For example, the ASCII "ADS Information" becomes hex "41445320496e666f726d6174696f6e.

scte35SegmentationDescriptor_segmentNum :: Lens' Scte35SegmentationDescriptor (Maybe Natural)

Corresponds to SCTE-35 segment_num. A value that is valid for the specified segmentation_type_id.

scte35SegmentationDescriptor_subSegmentsExpected :: Lens' Scte35SegmentationDescriptor (Maybe Natural)

Corresponds to SCTE-35 sub_segments_expected. A value that is valid for the specified segmentation_type_id.

scte35SegmentationDescriptor_segmentationUpidType :: Lens' Scte35SegmentationDescriptor (Maybe Natural)

Corresponds to SCTE-35 segmentation_upid_type. On the console, enter one of the types listed in the SCTE-35 specification, converted to a decimal. For example, "0x0C" hex from the specification is "12" in decimal. In the CLI, API, or an SDK, enter one of the types listed in the SCTE-35 specification, in either hex (for example, "0x0C" ) or in decimal (for example, "12").

scte35SegmentationDescriptor_segmentsExpected :: Lens' Scte35SegmentationDescriptor (Maybe Natural)

Corresponds to SCTE-35 segments_expected. A value that is valid for the specified segmentation_type_id.

scte35SegmentationDescriptor_segmentationTypeId :: Lens' Scte35SegmentationDescriptor (Maybe Natural)

Corresponds to SCTE-35 segmentation_type_id. One of the segmentation_type_id values listed in the SCTE-35 specification. On the console, enter the ID in decimal (for example, "52"). In the CLI, API, or an SDK, enter the ID in hex (for example, "0x34") or decimal (for example, "52").

scte35SegmentationDescriptor_deliveryRestrictions :: Lens' Scte35SegmentationDescriptor (Maybe Scte35DeliveryRestrictions)

Holds the four SCTE-35 delivery restriction parameters.

scte35SegmentationDescriptor_segmentationDuration :: Lens' Scte35SegmentationDescriptor (Maybe Natural)

Corresponds to SCTE-35 segmentation_duration. Optional. The duration for the time_signal, in 90 KHz ticks. To convert seconds to ticks, multiple the seconds by 90,000. Enter time in 90 KHz clock ticks. If you do not enter a duration, the time_signal will continue until you insert a cancellation message.

scte35SegmentationDescriptor_segmentationEventId :: Lens' Scte35SegmentationDescriptor Natural

Corresponds to SCTE-35 segmentation_event_id.

scte35SegmentationDescriptor_segmentationCancelIndicator :: Lens' Scte35SegmentationDescriptor Scte35SegmentationCancelIndicator

Corresponds to SCTE-35 segmentation_event_cancel_indicator.