Network.AWS.MediaLive.Types.Scte35SpliceInsert

Description

Documentation

data Scte35SpliceInsert

Scte35 Splice Insert

See: newScte35SpliceInsert smart constructor.

Constructors

Scte35SpliceInsert'

Fields

webDeliveryAllowedFlag :: Maybe Scte35SpliceInsertWebDeliveryAllowedBehavior

When set to ignore, Segment Descriptors with webDeliveryAllowedFlag set to 0 will no longer trigger blackouts or Ad Avail slates

adAvailOffset :: Maybe Int

When specified, this offset (in milliseconds) is added to the input Ad Avail PTS time. This only applies to embedded SCTE 104/35 messages and does not apply to OOB messages.

noRegionalBlackoutFlag :: Maybe Scte35SpliceInsertNoRegionalBlackoutBehavior

When set to ignore, Segment Descriptors with noRegionalBlackoutFlag set to 0 will no longer trigger blackouts or Ad Avail slates

newScte35SpliceInsert :: Scte35SpliceInsert

Create a value of Scte35SpliceInsert with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:webDeliveryAllowedFlag:Scte35SpliceInsert', scte35SpliceInsert_webDeliveryAllowedFlag - When set to ignore, Segment Descriptors with webDeliveryAllowedFlag set to 0 will no longer trigger blackouts or Ad Avail slates

$sel:adAvailOffset:Scte35SpliceInsert', scte35SpliceInsert_adAvailOffset - When specified, this offset (in milliseconds) is added to the input Ad Avail PTS time. This only applies to embedded SCTE 104/35 messages and does not apply to OOB messages.

$sel:noRegionalBlackoutFlag:Scte35SpliceInsert', scte35SpliceInsert_noRegionalBlackoutFlag - When set to ignore, Segment Descriptors with noRegionalBlackoutFlag set to 0 will no longer trigger blackouts or Ad Avail slates

scte35SpliceInsert_webDeliveryAllowedFlag :: Lens' Scte35SpliceInsert (Maybe Scte35SpliceInsertWebDeliveryAllowedBehavior)

When set to ignore, Segment Descriptors with webDeliveryAllowedFlag set to 0 will no longer trigger blackouts or Ad Avail slates

scte35SpliceInsert_adAvailOffset :: Lens' Scte35SpliceInsert (Maybe Int)

When specified, this offset (in milliseconds) is added to the input Ad Avail PTS time. This only applies to embedded SCTE 104/35 messages and does not apply to OOB messages.

scte35SpliceInsert_noRegionalBlackoutFlag :: Lens' Scte35SpliceInsert (Maybe Scte35SpliceInsertNoRegionalBlackoutBehavior)

When set to ignore, Segment Descriptors with noRegionalBlackoutFlag set to 0 will no longer trigger blackouts or Ad Avail slates