Network.AWS.MediaLive.Types.StaticImageActivateScheduleActionSettings

Description

Documentation

data StaticImageActivateScheduleActionSettings

Settings for the action to activate a static image.

See: newStaticImageActivateScheduleActionSettings smart constructor.

Constructors

StaticImageActivateScheduleActionSettings'

Fields

height :: Maybe Natural

The height of the image when inserted into the video, in pixels. The overlay will be scaled up or down to the specified height. Leave blank to use the native height of the overlay.

imageX :: Maybe Natural

Placement of the left edge of the overlay relative to the left edge of the video frame, in pixels. 0 (the default) is the left edge of the frame. If the placement causes the overlay to extend beyond the right edge of the underlying video, then the overlay is cropped on the right.

width :: Maybe Natural

The width of the image when inserted into the video, in pixels. The overlay will be scaled up or down to the specified width. Leave blank to use the native width of the overlay.

imageY :: Maybe Natural

Placement of the top edge of the overlay relative to the top edge of the video frame, in pixels. 0 (the default) is the top edge of the frame. If the placement causes the overlay to extend beyond the bottom edge of the underlying video, then the overlay is cropped on the bottom.

duration :: Maybe Natural

The duration in milliseconds for the image to remain on the video. If omitted or set to 0 the duration is unlimited and the image will remain until it is explicitly deactivated.

layer :: Maybe Natural

The number of the layer, 0 to 7. There are 8 layers that can be overlaid on the video, each layer with a different image. The layers are in Z order, which means that overlays with higher values of layer are inserted on top of overlays with lower values of layer. Default is 0.

opacity :: Maybe Natural

Opacity of image where 0 is transparent and 100 is fully opaque. Default is 100.

fadeIn :: Maybe Natural

The time in milliseconds for the image to fade in. The fade-in starts at the start time of the overlay. Default is 0 (no fade-in).

fadeOut :: Maybe Natural

Applies only if a duration is specified. The time in milliseconds for the image to fade out. The fade-out starts when the duration time is hit, so it effectively extends the duration. Default is 0 (no fade-out).

image :: InputLocation

The location and filename of the image file to overlay on the video. The file must be a 32-bit BMP, PNG, or TGA file, and must not be larger (in pixels) than the input video.

newStaticImageActivateScheduleActionSettings

Arguments

:: InputLocation
$sel:image:StaticImageActivateScheduleActionSettings'

-> StaticImageActivateScheduleActionSettings

Create a value of StaticImageActivateScheduleActionSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:height:StaticImageActivateScheduleActionSettings', staticImageActivateScheduleActionSettings_height - The height of the image when inserted into the video, in pixels. The overlay will be scaled up or down to the specified height. Leave blank to use the native height of the overlay.

$sel:imageX:StaticImageActivateScheduleActionSettings', staticImageActivateScheduleActionSettings_imageX - Placement of the left edge of the overlay relative to the left edge of the video frame, in pixels. 0 (the default) is the left edge of the frame. If the placement causes the overlay to extend beyond the right edge of the underlying video, then the overlay is cropped on the right.

$sel:width:StaticImageActivateScheduleActionSettings', staticImageActivateScheduleActionSettings_width - The width of the image when inserted into the video, in pixels. The overlay will be scaled up or down to the specified width. Leave blank to use the native width of the overlay.

$sel:imageY:StaticImageActivateScheduleActionSettings', staticImageActivateScheduleActionSettings_imageY - Placement of the top edge of the overlay relative to the top edge of the video frame, in pixels. 0 (the default) is the top edge of the frame. If the placement causes the overlay to extend beyond the bottom edge of the underlying video, then the overlay is cropped on the bottom.

$sel:duration:StaticImageActivateScheduleActionSettings', staticImageActivateScheduleActionSettings_duration - The duration in milliseconds for the image to remain on the video. If omitted or set to 0 the duration is unlimited and the image will remain until it is explicitly deactivated.

$sel:layer:StaticImageActivateScheduleActionSettings', staticImageActivateScheduleActionSettings_layer - The number of the layer, 0 to 7. There are 8 layers that can be overlaid on the video, each layer with a different image. The layers are in Z order, which means that overlays with higher values of layer are inserted on top of overlays with lower values of layer. Default is 0.

$sel:opacity:StaticImageActivateScheduleActionSettings', staticImageActivateScheduleActionSettings_opacity - Opacity of image where 0 is transparent and 100 is fully opaque. Default is 100.

$sel:fadeIn:StaticImageActivateScheduleActionSettings', staticImageActivateScheduleActionSettings_fadeIn - The time in milliseconds for the image to fade in. The fade-in starts at the start time of the overlay. Default is 0 (no fade-in).

$sel:fadeOut:StaticImageActivateScheduleActionSettings', staticImageActivateScheduleActionSettings_fadeOut - Applies only if a duration is specified. The time in milliseconds for the image to fade out. The fade-out starts when the duration time is hit, so it effectively extends the duration. Default is 0 (no fade-out).

$sel:image:StaticImageActivateScheduleActionSettings', staticImageActivateScheduleActionSettings_image - The location and filename of the image file to overlay on the video. The file must be a 32-bit BMP, PNG, or TGA file, and must not be larger (in pixels) than the input video.

staticImageActivateScheduleActionSettings_height :: Lens' StaticImageActivateScheduleActionSettings (Maybe Natural)

The height of the image when inserted into the video, in pixels. The overlay will be scaled up or down to the specified height. Leave blank to use the native height of the overlay.

staticImageActivateScheduleActionSettings_imageX :: Lens' StaticImageActivateScheduleActionSettings (Maybe Natural)

Placement of the left edge of the overlay relative to the left edge of the video frame, in pixels. 0 (the default) is the left edge of the frame. If the placement causes the overlay to extend beyond the right edge of the underlying video, then the overlay is cropped on the right.

staticImageActivateScheduleActionSettings_width :: Lens' StaticImageActivateScheduleActionSettings (Maybe Natural)

The width of the image when inserted into the video, in pixels. The overlay will be scaled up or down to the specified width. Leave blank to use the native width of the overlay.

staticImageActivateScheduleActionSettings_imageY :: Lens' StaticImageActivateScheduleActionSettings (Maybe Natural)

Placement of the top edge of the overlay relative to the top edge of the video frame, in pixels. 0 (the default) is the top edge of the frame. If the placement causes the overlay to extend beyond the bottom edge of the underlying video, then the overlay is cropped on the bottom.

staticImageActivateScheduleActionSettings_duration :: Lens' StaticImageActivateScheduleActionSettings (Maybe Natural)

The duration in milliseconds for the image to remain on the video. If omitted or set to 0 the duration is unlimited and the image will remain until it is explicitly deactivated.

staticImageActivateScheduleActionSettings_layer :: Lens' StaticImageActivateScheduleActionSettings (Maybe Natural)

The number of the layer, 0 to 7. There are 8 layers that can be overlaid on the video, each layer with a different image. The layers are in Z order, which means that overlays with higher values of layer are inserted on top of overlays with lower values of layer. Default is 0.

staticImageActivateScheduleActionSettings_opacity :: Lens' StaticImageActivateScheduleActionSettings (Maybe Natural)

Opacity of image where 0 is transparent and 100 is fully opaque. Default is 100.

staticImageActivateScheduleActionSettings_fadeIn :: Lens' StaticImageActivateScheduleActionSettings (Maybe Natural)

The time in milliseconds for the image to fade in. The fade-in starts at the start time of the overlay. Default is 0 (no fade-in).

staticImageActivateScheduleActionSettings_fadeOut :: Lens' StaticImageActivateScheduleActionSettings (Maybe Natural)

Applies only if a duration is specified. The time in milliseconds for the image to fade out. The fade-out starts when the duration time is hit, so it effectively extends the duration. Default is 0 (no fade-out).

staticImageActivateScheduleActionSettings_image :: Lens' StaticImageActivateScheduleActionSettings InputLocation

The location and filename of the image file to overlay on the video. The file must be a 32-bit BMP, PNG, or TGA file, and must not be larger (in pixels) than the input video.