Network.AWS.MediaLive.Types.TtmlDestinationSettings

Description

Documentation

data TtmlDestinationSettings

Ttml Destination Settings

See: newTtmlDestinationSettings smart constructor.

Constructors

TtmlDestinationSettings'

Fields

styleControl :: Maybe TtmlDestinationStyleControl

When set to passthrough, passes through style and position information from a TTML-like input source (TTML, SMPTE-TT, CFF-TT) to the CFF-TT output or TTML output.

newTtmlDestinationSettings :: TtmlDestinationSettings

Create a value of TtmlDestinationSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:styleControl:TtmlDestinationSettings', ttmlDestinationSettings_styleControl - When set to passthrough, passes through style and position information from a TTML-like input source (TTML, SMPTE-TT, CFF-TT) to the CFF-TT output or TTML output.

ttmlDestinationSettings_styleControl :: Lens' TtmlDestinationSettings (Maybe TtmlDestinationStyleControl)

When set to passthrough, passes through style and position information from a TTML-like input source (TTML, SMPTE-TT, CFF-TT) to the CFF-TT output or TTML output.