Network.AWS.MediaLive.Types.UdpOutputSettings

Description

Documentation

data UdpOutputSettings

Udp Output Settings

See: newUdpOutputSettings smart constructor.

Constructors

UdpOutputSettings'

Fields

bufferMsec :: Maybe Natural

UDP output buffering in milliseconds. Larger values increase latency through the transcoder but simultaneously assist the transcoder in maintaining a constant, low-jitter UDP/RTP output while accommodating clock recovery, input switching, input disruptions, picture reordering, etc.

fecOutputSettings :: Maybe FecOutputSettings

Settings for enabling and adjusting Forward Error Correction on UDP outputs.

destination :: OutputLocationRef

Destination address and port number for RTP or UDP packets. Can be unicast or multicast RTP or UDP (eg. rtp://239.10.10.10:5001 or udp://10.100.100.100:5002).

containerSettings :: UdpContainerSettings

newUdpOutputSettings

Arguments

:: OutputLocationRef
$sel:destination:UdpOutputSettings'

-> UdpContainerSettings
$sel:containerSettings:UdpOutputSettings'

-> UdpOutputSettings

Create a value of UdpOutputSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bufferMsec:UdpOutputSettings', udpOutputSettings_bufferMsec - UDP output buffering in milliseconds. Larger values increase latency through the transcoder but simultaneously assist the transcoder in maintaining a constant, low-jitter UDP/RTP output while accommodating clock recovery, input switching, input disruptions, picture reordering, etc.

$sel:fecOutputSettings:UdpOutputSettings', udpOutputSettings_fecOutputSettings - Settings for enabling and adjusting Forward Error Correction on UDP outputs.

$sel:destination:UdpOutputSettings', udpOutputSettings_destination - Destination address and port number for RTP or UDP packets. Can be unicast or multicast RTP or UDP (eg. rtp://239.10.10.10:5001 or udp://10.100.100.100:5002).

$sel:containerSettings:UdpOutputSettings', udpOutputSettings_containerSettings - Undocumented member.

udpOutputSettings_bufferMsec :: Lens' UdpOutputSettings (Maybe Natural)

UDP output buffering in milliseconds. Larger values increase latency through the transcoder but simultaneously assist the transcoder in maintaining a constant, low-jitter UDP/RTP output while accommodating clock recovery, input switching, input disruptions, picture reordering, etc.

udpOutputSettings_fecOutputSettings :: Lens' UdpOutputSettings (Maybe FecOutputSettings)

Settings for enabling and adjusting Forward Error Correction on UDP outputs.

udpOutputSettings_destination :: Lens' UdpOutputSettings OutputLocationRef

Destination address and port number for RTP or UDP packets. Can be unicast or multicast RTP or UDP (eg. rtp://239.10.10.10:5001 or udp://10.100.100.100:5002).

udpOutputSettings_containerSettings :: Lens' UdpOutputSettings UdpContainerSettings

Undocumented member.