Network.AWS.MediaLive.Types.VideoBlackFailoverSettings

Description

Documentation

data VideoBlackFailoverSettings

Placeholder documentation for VideoBlackFailoverSettings

See: newVideoBlackFailoverSettings smart constructor.

Constructors

VideoBlackFailoverSettings'

Fields

blackDetectThreshold :: Maybe Double

A value used in calculating the threshold below which MediaLive considers a pixel to be 'black'. For the input to be considered black, every pixel in a frame must be below this threshold. The threshold is calculated as a percentage (expressed as a decimal) of white. Therefore .1 means 10% white (or 90% black). Note how the formula works for any color depth. For example, if you set this field to 0.1 in 10-bit color depth: (1023*0.1=102.3), which means a pixel value of 102 or less is 'black'. If you set this field to .1 in an 8-bit color depth: (255*0.1=25.5), which means a pixel value of 25 or less is 'black'. The range is 0.0 to 1.0, with any number of decimal places.

videoBlackThresholdMsec :: Maybe Natural

The amount of time (in milliseconds) that the active input must be black before automatic input failover occurs.

newVideoBlackFailoverSettings :: VideoBlackFailoverSettings

Create a value of VideoBlackFailoverSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:blackDetectThreshold:VideoBlackFailoverSettings', videoBlackFailoverSettings_blackDetectThreshold - A value used in calculating the threshold below which MediaLive considers a pixel to be 'black'. For the input to be considered black, every pixel in a frame must be below this threshold. The threshold is calculated as a percentage (expressed as a decimal) of white. Therefore .1 means 10% white (or 90% black). Note how the formula works for any color depth. For example, if you set this field to 0.1 in 10-bit color depth: (1023*0.1=102.3), which means a pixel value of 102 or less is 'black'. If you set this field to .1 in an 8-bit color depth: (255*0.1=25.5), which means a pixel value of 25 or less is 'black'. The range is 0.0 to 1.0, with any number of decimal places.

$sel:videoBlackThresholdMsec:VideoBlackFailoverSettings', videoBlackFailoverSettings_videoBlackThresholdMsec - The amount of time (in milliseconds) that the active input must be black before automatic input failover occurs.

videoBlackFailoverSettings_blackDetectThreshold :: Lens' VideoBlackFailoverSettings (Maybe Double)

A value used in calculating the threshold below which MediaLive considers a pixel to be 'black'. For the input to be considered black, every pixel in a frame must be below this threshold. The threshold is calculated as a percentage (expressed as a decimal) of white. Therefore .1 means 10% white (or 90% black). Note how the formula works for any color depth. For example, if you set this field to 0.1 in 10-bit color depth: (1023*0.1=102.3), which means a pixel value of 102 or less is 'black'. If you set this field to .1 in an 8-bit color depth: (255*0.1=25.5), which means a pixel value of 25 or less is 'black'. The range is 0.0 to 1.0, with any number of decimal places.

videoBlackFailoverSettings_videoBlackThresholdMsec :: Lens' VideoBlackFailoverSettings (Maybe Natural)

The amount of time (in milliseconds) that the active input must be black before automatic input failover occurs.