Network.AWS.MediaLive.Types.VideoSelector

Description

Documentation

data VideoSelector

Specifies a particular video stream within an input source. An input may have only a single video selector.

See: newVideoSelector smart constructor.

Constructors

VideoSelector'

Fields

selectorSettings :: Maybe VideoSelectorSettings

The video selector settings.

colorSpaceUsage :: Maybe VideoSelectorColorSpaceUsage

Applies only if colorSpace is a value other than follow. This field controls how the value in the colorSpace field will be used. fallback means that when the input does include color space data, that data will be used, but when the input has no color space data, the value in colorSpace will be used. Choose fallback if your input is sometimes missing color space data, but when it does have color space data, that data is correct. force means to always use the value in colorSpace. Choose force if your input usually has no color space data or might have unreliable color space data.

colorSpaceSettings :: Maybe VideoSelectorColorSpaceSettings

Color space settings

colorSpace :: Maybe VideoSelectorColorSpace

Specifies the color space of an input. This setting works in tandem with colorSpaceUsage and a video description's colorSpaceSettingsChoice to determine if any conversion will be performed.

newVideoSelector :: VideoSelector

Create a value of VideoSelector with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:selectorSettings:VideoSelector', videoSelector_selectorSettings - The video selector settings.

$sel:colorSpaceUsage:VideoSelector', videoSelector_colorSpaceUsage - Applies only if colorSpace is a value other than follow. This field controls how the value in the colorSpace field will be used. fallback means that when the input does include color space data, that data will be used, but when the input has no color space data, the value in colorSpace will be used. Choose fallback if your input is sometimes missing color space data, but when it does have color space data, that data is correct. force means to always use the value in colorSpace. Choose force if your input usually has no color space data or might have unreliable color space data.

$sel:colorSpaceSettings:VideoSelector', videoSelector_colorSpaceSettings - Color space settings

$sel:colorSpace:VideoSelector', videoSelector_colorSpace - Specifies the color space of an input. This setting works in tandem with colorSpaceUsage and a video description's colorSpaceSettingsChoice to determine if any conversion will be performed.

videoSelector_selectorSettings :: Lens' VideoSelector (Maybe VideoSelectorSettings)

The video selector settings.

videoSelector_colorSpaceUsage :: Lens' VideoSelector (Maybe VideoSelectorColorSpaceUsage)

Applies only if colorSpace is a value other than follow. This field controls how the value in the colorSpace field will be used. fallback means that when the input does include color space data, that data will be used, but when the input has no color space data, the value in colorSpace will be used. Choose fallback if your input is sometimes missing color space data, but when it does have color space data, that data is correct. force means to always use the value in colorSpace. Choose force if your input usually has no color space data or might have unreliable color space data.

videoSelector_colorSpaceSettings :: Lens' VideoSelector (Maybe VideoSelectorColorSpaceSettings)

Color space settings

videoSelector_colorSpace :: Lens' VideoSelector (Maybe VideoSelectorColorSpace)

Specifies the color space of an input. This setting works in tandem with colorSpaceUsage and a video description's colorSpaceSettingsChoice to determine if any conversion will be performed.