Network.AWS.MediaLive.Types.VpcOutputSettings

Description

Documentation

data VpcOutputSettings

The properties for a private VPC Output When this property is specified, the output egress addresses will be created in a user specified VPC

See: newVpcOutputSettings smart constructor.

Constructors

VpcOutputSettings'

Fields

securityGroupIds :: Maybe [Text]

A list of up to 5 EC2 VPC security group IDs to attach to the Output VPC network interfaces. If none are specified then the VPC default security group will be used

publicAddressAllocationIds :: Maybe [Text]

List of public address allocation ids to associate with ENIs that will be created in Output VPC. Must specify one for SINGLE_PIPELINE, two for STANDARD channels

subnetIds :: [Text]

A list of VPC subnet IDs from the same VPC. If STANDARD channel, subnet IDs must be mapped to two unique availability zones (AZ).

newVpcOutputSettings :: VpcOutputSettings

Create a value of VpcOutputSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:securityGroupIds:VpcOutputSettings', vpcOutputSettings_securityGroupIds - A list of up to 5 EC2 VPC security group IDs to attach to the Output VPC network interfaces. If none are specified then the VPC default security group will be used

$sel:publicAddressAllocationIds:VpcOutputSettings', vpcOutputSettings_publicAddressAllocationIds - List of public address allocation ids to associate with ENIs that will be created in Output VPC. Must specify one for SINGLE_PIPELINE, two for STANDARD channels

$sel:subnetIds:VpcOutputSettings', vpcOutputSettings_subnetIds - A list of VPC subnet IDs from the same VPC. If STANDARD channel, subnet IDs must be mapped to two unique availability zones (AZ).

vpcOutputSettings_securityGroupIds :: Lens' VpcOutputSettings (Maybe [Text])

A list of up to 5 EC2 VPC security group IDs to attach to the Output VPC network interfaces. If none are specified then the VPC default security group will be used

vpcOutputSettings_publicAddressAllocationIds :: Lens' VpcOutputSettings (Maybe [Text])

List of public address allocation ids to associate with ENIs that will be created in Output VPC. Must specify one for SINGLE_PIPELINE, two for STANDARD channels

vpcOutputSettings_subnetIds :: Lens' VpcOutputSettings [Text]

A list of VPC subnet IDs from the same VPC. If STANDARD channel, subnet IDs must be mapped to two unique availability zones (AZ).