Network.AWS.MediaLive.UpdateChannel

Description

Updates a channel.

Creating a Request

data UpdateChannel'

A request to update a channel.

See: newUpdateChannel' smart constructor.

Constructors

UpdateChannel''

Fields

logLevel :: Maybe LogLevel

The log level to write to CloudWatch Logs.

inputSpecification :: Maybe InputSpecification

Specification of network and file inputs for this channel

inputAttachments :: Maybe [InputAttachment]

destinations :: Maybe [OutputDestination]

A list of output destinations for this channel.

name :: Maybe Text

The name of the channel.

cdiInputSpecification :: Maybe CdiInputSpecification

Specification of CDI inputs for this channel

encoderSettings :: Maybe EncoderSettings

The encoder settings for this channel.

roleArn :: Maybe Text

An optional Amazon Resource Name (ARN) of the role to assume when running the Channel. If you do not specify this on an update call but the role was previously set that role will be removed.

channelId :: Text

channel ID

newUpdateChannel'

Arguments

:: Text
$sel:channelId:UpdateChannel''

-> UpdateChannel'

Create a value of UpdateChannel' with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logLevel:UpdateChannel'', updateChannel'_logLevel - The log level to write to CloudWatch Logs.

$sel:inputSpecification:UpdateChannel'', updateChannel'_inputSpecification - Specification of network and file inputs for this channel

$sel:inputAttachments:UpdateChannel'', updateChannel'_inputAttachments - Undocumented member.

$sel:destinations:UpdateChannel'', updateChannel'_destinations - A list of output destinations for this channel.

$sel:name:UpdateChannel'', updateChannel'_name - The name of the channel.

$sel:cdiInputSpecification:UpdateChannel'', updateChannel'_cdiInputSpecification - Specification of CDI inputs for this channel

$sel:encoderSettings:UpdateChannel'', updateChannel'_encoderSettings - The encoder settings for this channel.

$sel:roleArn:UpdateChannel'', updateChannel'_roleArn - An optional Amazon Resource Name (ARN) of the role to assume when running the Channel. If you do not specify this on an update call but the role was previously set that role will be removed.

$sel:channelId:UpdateChannel'', updateChannel'_channelId - channel ID

Request Lenses

updateChannel'_logLevel :: Lens' UpdateChannel' (Maybe LogLevel)

The log level to write to CloudWatch Logs.

updateChannel'_inputSpecification :: Lens' UpdateChannel' (Maybe InputSpecification)

Specification of network and file inputs for this channel

updateChannel'_inputAttachments :: Lens' UpdateChannel' (Maybe [InputAttachment])

Undocumented member.

updateChannel'_destinations :: Lens' UpdateChannel' (Maybe [OutputDestination])

A list of output destinations for this channel.

updateChannel'_name :: Lens' UpdateChannel' (Maybe Text)

The name of the channel.

updateChannel'_cdiInputSpecification :: Lens' UpdateChannel' (Maybe CdiInputSpecification)

Specification of CDI inputs for this channel

updateChannel'_encoderSettings :: Lens' UpdateChannel' (Maybe EncoderSettings)

The encoder settings for this channel.

updateChannel'_roleArn :: Lens' UpdateChannel' (Maybe Text)

An optional Amazon Resource Name (ARN) of the role to assume when running the Channel. If you do not specify this on an update call but the role was previously set that role will be removed.

updateChannel'_channelId :: Lens' UpdateChannel' Text

channel ID

Destructuring the Response

data UpdateChannelResponse

Placeholder documentation for UpdateChannelResponse

See: newUpdateChannelResponse smart constructor.

Constructors

UpdateChannelResponse'

Fields

channel :: Maybe Channel

httpStatus :: Int

The response's http status code.

newUpdateChannelResponse

Arguments

:: Int
$sel:httpStatus:UpdateChannelResponse'

-> UpdateChannelResponse

Create a value of UpdateChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channel:UpdateChannelResponse', updateChannelResponse_channel - Undocumented member.

$sel:httpStatus:UpdateChannelResponse', updateChannelResponse_httpStatus - The response's http status code.

Response Lenses

updateChannelResponse_channel :: Lens' UpdateChannelResponse (Maybe Channel)

Undocumented member.

updateChannelResponse_httpStatus :: Lens' UpdateChannelResponse Int

The response's http status code.