Network.AWS.MediaLive.UpdateInputDevice

Description

Updates the parameters for the input device.

Creating a Request

data UpdateInputDevice'

A request to update an input device.

See: newUpdateInputDevice' smart constructor.

Constructors

UpdateInputDevice''

Fields

hdDeviceSettings :: Maybe InputDeviceConfigurableSettings

The settings that you want to apply to the HD input device.

uhdDeviceSettings :: Maybe InputDeviceConfigurableSettings

The settings that you want to apply to the UHD input device.

name :: Maybe Text

The name that you assigned to this input device (not the unique ID).

inputDeviceId :: Text

The unique ID of the input device. For example, hd-123456789abcdef.

newUpdateInputDevice'

Arguments

:: Text
$sel:inputDeviceId:UpdateInputDevice''

-> UpdateInputDevice'

Create a value of UpdateInputDevice' with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hdDeviceSettings:UpdateInputDevice'', updateInputDevice'_hdDeviceSettings - The settings that you want to apply to the HD input device.

$sel:uhdDeviceSettings:UpdateInputDevice'', updateInputDevice'_uhdDeviceSettings - The settings that you want to apply to the UHD input device.

$sel:name:UpdateInputDevice'', updateInputDevice'_name - The name that you assigned to this input device (not the unique ID).

$sel:inputDeviceId:UpdateInputDevice'', updateInputDevice'_inputDeviceId - The unique ID of the input device. For example, hd-123456789abcdef.

Request Lenses

updateInputDevice'_hdDeviceSettings :: Lens' UpdateInputDevice' (Maybe InputDeviceConfigurableSettings)

The settings that you want to apply to the HD input device.

updateInputDevice'_uhdDeviceSettings :: Lens' UpdateInputDevice' (Maybe InputDeviceConfigurableSettings)

The settings that you want to apply to the UHD input device.

updateInputDevice'_name :: Lens' UpdateInputDevice' (Maybe Text)

The name that you assigned to this input device (not the unique ID).

updateInputDevice'_inputDeviceId :: Lens' UpdateInputDevice' Text

The unique ID of the input device. For example, hd-123456789abcdef.

Destructuring the Response

data UpdateInputDeviceResponse

Placeholder documentation for UpdateInputDeviceResponse

See: newUpdateInputDeviceResponse smart constructor.

Constructors

UpdateInputDeviceResponse'

Fields

arn :: Maybe Text

The unique ARN of the input device.

macAddress :: Maybe Text

The network MAC address of the input device.

hdDeviceSettings :: Maybe InputDeviceHdSettings

Settings that describe an input device that is type HD.

uhdDeviceSettings :: Maybe InputDeviceUhdSettings

Settings that describe an input device that is type UHD.

name :: Maybe Text

A name that you specify for the input device.

id :: Maybe Text

The unique ID of the input device.

deviceUpdateStatus :: Maybe DeviceUpdateStatus

The status of software on the input device.

deviceSettingsSyncState :: Maybe DeviceSettingsSyncState

The status of the action to synchronize the device configuration. If you change the configuration of the input device (for example, the maximum bitrate), MediaLive sends the new data to the device. The device might not update itself immediately. SYNCED means the device has updated its configuration. SYNCING means that it has not updated its configuration.

type' :: Maybe InputDeviceType

The type of the input device.

serialNumber :: Maybe Text

The unique serial number of the input device.

networkSettings :: Maybe InputDeviceNetworkSettings

The network settings for the input device.

connectionState :: Maybe InputDeviceConnectionState

The state of the connection between the input device and AWS.

httpStatus :: Int

The response's http status code.

newUpdateInputDeviceResponse

Arguments

:: Int
$sel:httpStatus:UpdateInputDeviceResponse'

-> UpdateInputDeviceResponse

Create a value of UpdateInputDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:UpdateInputDeviceResponse', updateInputDeviceResponse_arn - The unique ARN of the input device.

$sel:macAddress:UpdateInputDeviceResponse', updateInputDeviceResponse_macAddress - The network MAC address of the input device.

$sel:hdDeviceSettings:UpdateInputDevice'', updateInputDeviceResponse_hdDeviceSettings - Settings that describe an input device that is type HD.

$sel:uhdDeviceSettings:UpdateInputDevice'', updateInputDeviceResponse_uhdDeviceSettings - Settings that describe an input device that is type UHD.

$sel:name:UpdateInputDevice'', updateInputDeviceResponse_name - A name that you specify for the input device.

$sel:id:UpdateInputDeviceResponse', updateInputDeviceResponse_id - The unique ID of the input device.

$sel:deviceUpdateStatus:UpdateInputDeviceResponse', updateInputDeviceResponse_deviceUpdateStatus - The status of software on the input device.

$sel:deviceSettingsSyncState:UpdateInputDeviceResponse', updateInputDeviceResponse_deviceSettingsSyncState - The status of the action to synchronize the device configuration. If you change the configuration of the input device (for example, the maximum bitrate), MediaLive sends the new data to the device. The device might not update itself immediately. SYNCED means the device has updated its configuration. SYNCING means that it has not updated its configuration.

$sel:type':UpdateInputDeviceResponse', updateInputDeviceResponse_type - The type of the input device.

$sel:serialNumber:UpdateInputDeviceResponse', updateInputDeviceResponse_serialNumber - The unique serial number of the input device.

$sel:networkSettings:UpdateInputDeviceResponse', updateInputDeviceResponse_networkSettings - The network settings for the input device.

$sel:connectionState:UpdateInputDeviceResponse', updateInputDeviceResponse_connectionState - The state of the connection between the input device and AWS.

$sel:httpStatus:UpdateInputDeviceResponse', updateInputDeviceResponse_httpStatus - The response's http status code.

Response Lenses

updateInputDeviceResponse_arn :: Lens' UpdateInputDeviceResponse (Maybe Text)

The unique ARN of the input device.

updateInputDeviceResponse_macAddress :: Lens' UpdateInputDeviceResponse (Maybe Text)

The network MAC address of the input device.

updateInputDeviceResponse_hdDeviceSettings :: Lens' UpdateInputDeviceResponse (Maybe InputDeviceHdSettings)

Settings that describe an input device that is type HD.

updateInputDeviceResponse_uhdDeviceSettings :: Lens' UpdateInputDeviceResponse (Maybe InputDeviceUhdSettings)

Settings that describe an input device that is type UHD.

updateInputDeviceResponse_name :: Lens' UpdateInputDeviceResponse (Maybe Text)

A name that you specify for the input device.

updateInputDeviceResponse_id :: Lens' UpdateInputDeviceResponse (Maybe Text)

The unique ID of the input device.

updateInputDeviceResponse_deviceUpdateStatus :: Lens' UpdateInputDeviceResponse (Maybe DeviceUpdateStatus)

The status of software on the input device.

updateInputDeviceResponse_deviceSettingsSyncState :: Lens' UpdateInputDeviceResponse (Maybe DeviceSettingsSyncState)

The status of the action to synchronize the device configuration. If you change the configuration of the input device (for example, the maximum bitrate), MediaLive sends the new data to the device. The device might not update itself immediately. SYNCED means the device has updated its configuration. SYNCING means that it has not updated its configuration.

updateInputDeviceResponse_type :: Lens' UpdateInputDeviceResponse (Maybe InputDeviceType)

The type of the input device.

updateInputDeviceResponse_serialNumber :: Lens' UpdateInputDeviceResponse (Maybe Text)

The unique serial number of the input device.

updateInputDeviceResponse_networkSettings :: Lens' UpdateInputDeviceResponse (Maybe InputDeviceNetworkSettings)

The network settings for the input device.

updateInputDeviceResponse_connectionState :: Lens' UpdateInputDeviceResponse (Maybe InputDeviceConnectionState)

The state of the connection between the input device and AWS.

updateInputDeviceResponse_httpStatus :: Lens' UpdateInputDeviceResponse Int

The response's http status code.