Network.AWS.MediaPackage.CreateChannel

Description

Creates a new Channel.

Creating a Request

data CreateChannel

A new Channel configuration.

See: newCreateChannel smart constructor.

Constructors

CreateChannel'

Fields

tags :: Maybe (HashMap Text Text)

description :: Maybe Text

A short text description of the Channel.

id :: Text

The ID of the Channel. The ID must be unique within the region and it cannot be changed after a Channel is created.

newCreateChannel

Arguments

:: Text
$sel:id:CreateChannel'

-> CreateChannel

Create a value of CreateChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateChannel', createChannel_tags - Undocumented member.

$sel:description:CreateChannel', createChannel_description - A short text description of the Channel.

$sel:id:CreateChannel', createChannel_id - The ID of the Channel. The ID must be unique within the region and it cannot be changed after a Channel is created.

Request Lenses

createChannel_tags :: Lens' CreateChannel (Maybe (HashMap Text Text))

Undocumented member.

createChannel_description :: Lens' CreateChannel (Maybe Text)

A short text description of the Channel.

createChannel_id :: Lens' CreateChannel Text

The ID of the Channel. The ID must be unique within the region and it cannot be changed after a Channel is created.

Destructuring the Response

data CreateChannelResponse

See: newCreateChannelResponse smart constructor.

Constructors

CreateChannelResponse'

Fields

egressAccessLogs :: Maybe EgressAccessLogs

arn :: Maybe Text

The Amazon Resource Name (ARN) assigned to the Channel.

id :: Maybe Text

The ID of the Channel.

hlsIngest :: Maybe HlsIngest

ingressAccessLogs :: Maybe IngressAccessLogs

tags :: Maybe (HashMap Text Text)

description :: Maybe Text

A short text description of the Channel.

httpStatus :: Int

The response's http status code.

newCreateChannelResponse

Arguments

:: Int
$sel:httpStatus:CreateChannelResponse'

-> CreateChannelResponse

Create a value of CreateChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:egressAccessLogs:CreateChannelResponse', createChannelResponse_egressAccessLogs - Undocumented member.

$sel:arn:CreateChannelResponse', createChannelResponse_arn - The Amazon Resource Name (ARN) assigned to the Channel.

$sel:id:CreateChannel', createChannelResponse_id - The ID of the Channel.

$sel:hlsIngest:CreateChannelResponse', createChannelResponse_hlsIngest - Undocumented member.

$sel:ingressAccessLogs:CreateChannelResponse', createChannelResponse_ingressAccessLogs - Undocumented member.

$sel:tags:CreateChannel', createChannelResponse_tags - Undocumented member.

$sel:description:CreateChannel', createChannelResponse_description - A short text description of the Channel.

$sel:httpStatus:CreateChannelResponse', createChannelResponse_httpStatus - The response's http status code.

Response Lenses

createChannelResponse_egressAccessLogs :: Lens' CreateChannelResponse (Maybe EgressAccessLogs)

Undocumented member.

createChannelResponse_arn :: Lens' CreateChannelResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the Channel.

createChannelResponse_id :: Lens' CreateChannelResponse (Maybe Text)

The ID of the Channel.

createChannelResponse_hlsIngest :: Lens' CreateChannelResponse (Maybe HlsIngest)

Undocumented member.

createChannelResponse_ingressAccessLogs :: Lens' CreateChannelResponse (Maybe IngressAccessLogs)

Undocumented member.

createChannelResponse_tags :: Lens' CreateChannelResponse (Maybe (HashMap Text Text))

Undocumented member.

createChannelResponse_description :: Lens' CreateChannelResponse (Maybe Text)

A short text description of the Channel.

createChannelResponse_httpStatus :: Lens' CreateChannelResponse Int

The response's http status code.