Network.AWS.MediaPackage.CreateHarvestJob

Description

Creates a new HarvestJob record.

Creating a Request

data CreateHarvestJob

Configuration parameters used to create a new HarvestJob.

See: newCreateHarvestJob smart constructor.

Constructors

CreateHarvestJob'

Fields

s3Destination :: S3Destination

endTime :: Text

The end of the time-window which will be harvested

originEndpointId :: Text

The ID of the OriginEndpoint that the HarvestJob will harvest from. This cannot be changed after the HarvestJob is submitted.

startTime :: Text

The start of the time-window which will be harvested

id :: Text

The ID of the HarvestJob. The ID must be unique within the region and it cannot be changed after the HarvestJob is submitted

newCreateHarvestJob

Arguments

:: S3Destination
$sel:s3Destination:CreateHarvestJob'

-> Text
$sel:endTime:CreateHarvestJob'

-> Text
$sel:originEndpointId:CreateHarvestJob'

-> Text
$sel:startTime:CreateHarvestJob'

-> Text
$sel:id:CreateHarvestJob'

-> CreateHarvestJob

Create a value of CreateHarvestJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Destination:CreateHarvestJob', createHarvestJob_s3Destination - Undocumented member.

$sel:endTime:CreateHarvestJob', createHarvestJob_endTime - The end of the time-window which will be harvested

$sel:originEndpointId:CreateHarvestJob', createHarvestJob_originEndpointId - The ID of the OriginEndpoint that the HarvestJob will harvest from. This cannot be changed after the HarvestJob is submitted.

$sel:startTime:CreateHarvestJob', createHarvestJob_startTime - The start of the time-window which will be harvested

$sel:id:CreateHarvestJob', createHarvestJob_id - The ID of the HarvestJob. The ID must be unique within the region and it cannot be changed after the HarvestJob is submitted

Request Lenses

createHarvestJob_s3Destination :: Lens' CreateHarvestJob S3Destination

Undocumented member.

createHarvestJob_endTime :: Lens' CreateHarvestJob Text

The end of the time-window which will be harvested

createHarvestJob_originEndpointId :: Lens' CreateHarvestJob Text

The ID of the OriginEndpoint that the HarvestJob will harvest from. This cannot be changed after the HarvestJob is submitted.

createHarvestJob_startTime :: Lens' CreateHarvestJob Text

The start of the time-window which will be harvested

createHarvestJob_id :: Lens' CreateHarvestJob Text

The ID of the HarvestJob. The ID must be unique within the region and it cannot be changed after the HarvestJob is submitted

Destructuring the Response

data CreateHarvestJobResponse

See: newCreateHarvestJobResponse smart constructor.

Constructors

CreateHarvestJobResponse'

Fields

s3Destination :: Maybe S3Destination

status :: Maybe Status

The current status of the HarvestJob. Consider setting up a CloudWatch Event to listen for HarvestJobs as they succeed or fail. In the event of failure, the CloudWatch Event will include an explanation of why the HarvestJob failed.

channelId :: Maybe Text

The ID of the Channel that the HarvestJob will harvest from.

createdAt :: Maybe Text

The time the HarvestJob was submitted

arn :: Maybe Text

The Amazon Resource Name (ARN) assigned to the HarvestJob.

id :: Maybe Text

The ID of the HarvestJob. The ID must be unique within the region and it cannot be changed after the HarvestJob is submitted.

startTime :: Maybe Text

The start of the time-window which will be harvested.

endTime :: Maybe Text

The end of the time-window which will be harvested.

originEndpointId :: Maybe Text

The ID of the OriginEndpoint that the HarvestJob will harvest from. This cannot be changed after the HarvestJob is submitted.

httpStatus :: Int

The response's http status code.

newCreateHarvestJobResponse

Arguments

:: Int
$sel:httpStatus:CreateHarvestJobResponse'

-> CreateHarvestJobResponse

Create a value of CreateHarvestJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Destination:CreateHarvestJob', createHarvestJobResponse_s3Destination - Undocumented member.

$sel:status:CreateHarvestJobResponse', createHarvestJobResponse_status - The current status of the HarvestJob. Consider setting up a CloudWatch Event to listen for HarvestJobs as they succeed or fail. In the event of failure, the CloudWatch Event will include an explanation of why the HarvestJob failed.

$sel:channelId:CreateHarvestJobResponse', createHarvestJobResponse_channelId - The ID of the Channel that the HarvestJob will harvest from.

$sel:createdAt:CreateHarvestJobResponse', createHarvestJobResponse_createdAt - The time the HarvestJob was submitted

$sel:arn:CreateHarvestJobResponse', createHarvestJobResponse_arn - The Amazon Resource Name (ARN) assigned to the HarvestJob.

$sel:id:CreateHarvestJob', createHarvestJobResponse_id - The ID of the HarvestJob. The ID must be unique within the region and it cannot be changed after the HarvestJob is submitted.

$sel:startTime:CreateHarvestJob', createHarvestJobResponse_startTime - The start of the time-window which will be harvested.

$sel:endTime:CreateHarvestJob', createHarvestJobResponse_endTime - The end of the time-window which will be harvested.

$sel:originEndpointId:CreateHarvestJob', createHarvestJobResponse_originEndpointId - The ID of the OriginEndpoint that the HarvestJob will harvest from. This cannot be changed after the HarvestJob is submitted.

$sel:httpStatus:CreateHarvestJobResponse', createHarvestJobResponse_httpStatus - The response's http status code.

Response Lenses

createHarvestJobResponse_s3Destination :: Lens' CreateHarvestJobResponse (Maybe S3Destination)

Undocumented member.

createHarvestJobResponse_status :: Lens' CreateHarvestJobResponse (Maybe Status)

The current status of the HarvestJob. Consider setting up a CloudWatch Event to listen for HarvestJobs as they succeed or fail. In the event of failure, the CloudWatch Event will include an explanation of why the HarvestJob failed.

createHarvestJobResponse_channelId :: Lens' CreateHarvestJobResponse (Maybe Text)

The ID of the Channel that the HarvestJob will harvest from.

createHarvestJobResponse_createdAt :: Lens' CreateHarvestJobResponse (Maybe Text)

The time the HarvestJob was submitted

createHarvestJobResponse_arn :: Lens' CreateHarvestJobResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the HarvestJob.

createHarvestJobResponse_id :: Lens' CreateHarvestJobResponse (Maybe Text)

The ID of the HarvestJob. The ID must be unique within the region and it cannot be changed after the HarvestJob is submitted.

createHarvestJobResponse_startTime :: Lens' CreateHarvestJobResponse (Maybe Text)

The start of the time-window which will be harvested.

createHarvestJobResponse_endTime :: Lens' CreateHarvestJobResponse (Maybe Text)

The end of the time-window which will be harvested.

createHarvestJobResponse_originEndpointId :: Lens' CreateHarvestJobResponse (Maybe Text)

The ID of the OriginEndpoint that the HarvestJob will harvest from. This cannot be changed after the HarvestJob is submitted.

createHarvestJobResponse_httpStatus :: Lens' CreateHarvestJobResponse Int

The response's http status code.