Network.AWS.MediaPackage.DescribeOriginEndpoint

Description

Gets details about an existing OriginEndpoint.

Creating a Request

data DescribeOriginEndpoint

See: newDescribeOriginEndpoint smart constructor.

Constructors

DescribeOriginEndpoint'

Fields

id :: Text

The ID of the OriginEndpoint.

newDescribeOriginEndpoint

Arguments

:: Text
$sel:id:DescribeOriginEndpoint'

-> DescribeOriginEndpoint

Create a value of DescribeOriginEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DescribeOriginEndpoint', describeOriginEndpoint_id - The ID of the OriginEndpoint.

Request Lenses

describeOriginEndpoint_id :: Lens' DescribeOriginEndpoint Text

The ID of the OriginEndpoint.

Destructuring the Response

data DescribeOriginEndpointResponse

See: newDescribeOriginEndpointResponse smart constructor.

Constructors

DescribeOriginEndpointResponse'

Fields

dashPackage :: Maybe DashPackage

origination :: Maybe Origination

Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

startoverWindowSeconds :: Maybe Int

Maximum duration (seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

channelId :: Maybe Text

The ID of the Channel the OriginEndpoint is associated with.

cmafPackage :: Maybe CmafPackage

manifestName :: Maybe Text

A short string appended to the end of the OriginEndpoint URL.

arn :: Maybe Text

The Amazon Resource Name (ARN) assigned to the OriginEndpoint.

id :: Maybe Text

The ID of the OriginEndpoint.

whitelist :: Maybe [Text]

A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

mssPackage :: Maybe MssPackage

tags :: Maybe (HashMap Text Text)

description :: Maybe Text

A short text description of the OriginEndpoint.

url :: Maybe Text

The URL of the packaged OriginEndpoint for consumption.

timeDelaySeconds :: Maybe Int

Amount of delay (seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

authorization :: Maybe Authorization

hlsPackage :: Maybe HlsPackage

httpStatus :: Int

The response's http status code.

newDescribeOriginEndpointResponse

Arguments

:: Int
$sel:httpStatus:DescribeOriginEndpointResponse'

-> DescribeOriginEndpointResponse

Create a value of DescribeOriginEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dashPackage:DescribeOriginEndpointResponse', describeOriginEndpointResponse_dashPackage - Undocumented member.

$sel:origination:DescribeOriginEndpointResponse', describeOriginEndpointResponse_origination - Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

$sel:startoverWindowSeconds:DescribeOriginEndpointResponse', describeOriginEndpointResponse_startoverWindowSeconds - Maximum duration (seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

$sel:channelId:DescribeOriginEndpointResponse', describeOriginEndpointResponse_channelId - The ID of the Channel the OriginEndpoint is associated with.

$sel:cmafPackage:DescribeOriginEndpointResponse', describeOriginEndpointResponse_cmafPackage - Undocumented member.

$sel:manifestName:DescribeOriginEndpointResponse', describeOriginEndpointResponse_manifestName - A short string appended to the end of the OriginEndpoint URL.

$sel:arn:DescribeOriginEndpointResponse', describeOriginEndpointResponse_arn - The Amazon Resource Name (ARN) assigned to the OriginEndpoint.

$sel:id:DescribeOriginEndpoint', describeOriginEndpointResponse_id - The ID of the OriginEndpoint.

$sel:whitelist:DescribeOriginEndpointResponse', describeOriginEndpointResponse_whitelist - A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

$sel:mssPackage:DescribeOriginEndpointResponse', describeOriginEndpointResponse_mssPackage - Undocumented member.

$sel:tags:DescribeOriginEndpointResponse', describeOriginEndpointResponse_tags - Undocumented member.

$sel:description:DescribeOriginEndpointResponse', describeOriginEndpointResponse_description - A short text description of the OriginEndpoint.

$sel:url:DescribeOriginEndpointResponse', describeOriginEndpointResponse_url - The URL of the packaged OriginEndpoint for consumption.

$sel:timeDelaySeconds:DescribeOriginEndpointResponse', describeOriginEndpointResponse_timeDelaySeconds - Amount of delay (seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

$sel:authorization:DescribeOriginEndpointResponse', describeOriginEndpointResponse_authorization - Undocumented member.

$sel:hlsPackage:DescribeOriginEndpointResponse', describeOriginEndpointResponse_hlsPackage - Undocumented member.

$sel:httpStatus:DescribeOriginEndpointResponse', describeOriginEndpointResponse_httpStatus - The response's http status code.

Response Lenses

describeOriginEndpointResponse_dashPackage :: Lens' DescribeOriginEndpointResponse (Maybe DashPackage)

Undocumented member.

describeOriginEndpointResponse_origination :: Lens' DescribeOriginEndpointResponse (Maybe Origination)

Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

describeOriginEndpointResponse_startoverWindowSeconds :: Lens' DescribeOriginEndpointResponse (Maybe Int)

Maximum duration (seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

describeOriginEndpointResponse_channelId :: Lens' DescribeOriginEndpointResponse (Maybe Text)

The ID of the Channel the OriginEndpoint is associated with.

describeOriginEndpointResponse_cmafPackage :: Lens' DescribeOriginEndpointResponse (Maybe CmafPackage)

Undocumented member.

describeOriginEndpointResponse_manifestName :: Lens' DescribeOriginEndpointResponse (Maybe Text)

A short string appended to the end of the OriginEndpoint URL.

describeOriginEndpointResponse_arn :: Lens' DescribeOriginEndpointResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the OriginEndpoint.

describeOriginEndpointResponse_id :: Lens' DescribeOriginEndpointResponse (Maybe Text)

The ID of the OriginEndpoint.

describeOriginEndpointResponse_whitelist :: Lens' DescribeOriginEndpointResponse (Maybe [Text])

A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

describeOriginEndpointResponse_mssPackage :: Lens' DescribeOriginEndpointResponse (Maybe MssPackage)

Undocumented member.

describeOriginEndpointResponse_tags :: Lens' DescribeOriginEndpointResponse (Maybe (HashMap Text Text))

Undocumented member.

describeOriginEndpointResponse_description :: Lens' DescribeOriginEndpointResponse (Maybe Text)

A short text description of the OriginEndpoint.

describeOriginEndpointResponse_url :: Lens' DescribeOriginEndpointResponse (Maybe Text)

The URL of the packaged OriginEndpoint for consumption.

describeOriginEndpointResponse_timeDelaySeconds :: Lens' DescribeOriginEndpointResponse (Maybe Int)

Amount of delay (seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

describeOriginEndpointResponse_authorization :: Lens' DescribeOriginEndpointResponse (Maybe Authorization)

Undocumented member.

describeOriginEndpointResponse_hlsPackage :: Lens' DescribeOriginEndpointResponse (Maybe HlsPackage)

Undocumented member.

describeOriginEndpointResponse_httpStatus :: Lens' DescribeOriginEndpointResponse Int

The response's http status code.