Network.AWS.MediaPackage.Lens

Description

Operations

CreateHarvestJob

createHarvestJob_s3Destination :: Lens' CreateHarvestJob S3Destination

Undocumented member.

createHarvestJob_endTime :: Lens' CreateHarvestJob Text

The end of the time-window which will be harvested

createHarvestJob_originEndpointId :: Lens' CreateHarvestJob Text

The ID of the OriginEndpoint that the HarvestJob will harvest from. This cannot be changed after the HarvestJob is submitted.

createHarvestJob_startTime :: Lens' CreateHarvestJob Text

The start of the time-window which will be harvested

createHarvestJob_id :: Lens' CreateHarvestJob Text

The ID of the HarvestJob. The ID must be unique within the region and it cannot be changed after the HarvestJob is submitted

createHarvestJobResponse_status :: Lens' CreateHarvestJobResponse (Maybe Status)

The current status of the HarvestJob. Consider setting up a CloudWatch Event to listen for HarvestJobs as they succeed or fail. In the event of failure, the CloudWatch Event will include an explanation of why the HarvestJob failed.

createHarvestJobResponse_originEndpointId :: Lens' CreateHarvestJobResponse (Maybe Text)

The ID of the OriginEndpoint that the HarvestJob will harvest from. This cannot be changed after the HarvestJob is submitted.

createHarvestJobResponse_startTime :: Lens' CreateHarvestJobResponse (Maybe Text)

The start of the time-window which will be harvested.

createHarvestJobResponse_arn :: Lens' CreateHarvestJobResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the HarvestJob.

createHarvestJobResponse_createdAt :: Lens' CreateHarvestJobResponse (Maybe Text)

The time the HarvestJob was submitted

createHarvestJobResponse_channelId :: Lens' CreateHarvestJobResponse (Maybe Text)

The ID of the Channel that the HarvestJob will harvest from.

createHarvestJobResponse_s3Destination :: Lens' CreateHarvestJobResponse (Maybe S3Destination)

Undocumented member.

createHarvestJobResponse_endTime :: Lens' CreateHarvestJobResponse (Maybe Text)

The end of the time-window which will be harvested.

createHarvestJobResponse_id :: Lens' CreateHarvestJobResponse (Maybe Text)

The ID of the HarvestJob. The ID must be unique within the region and it cannot be changed after the HarvestJob is submitted.

createHarvestJobResponse_httpStatus :: Lens' CreateHarvestJobResponse Int

The response's http status code.

ConfigureLogs

configureLogs_ingressAccessLogs :: Lens' ConfigureLogs (Maybe IngressAccessLogs)

Undocumented member.

configureLogs_egressAccessLogs :: Lens' ConfigureLogs (Maybe EgressAccessLogs)

Undocumented member.

configureLogs_id :: Lens' ConfigureLogs Text

The ID of the channel to log subscription.

configureLogsResponse_ingressAccessLogs :: Lens' ConfigureLogsResponse (Maybe IngressAccessLogs)

Undocumented member.

configureLogsResponse_hlsIngest :: Lens' ConfigureLogsResponse (Maybe HlsIngest)

Undocumented member.

configureLogsResponse_arn :: Lens' ConfigureLogsResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the Channel.

configureLogsResponse_id :: Lens' ConfigureLogsResponse (Maybe Text)

The ID of the Channel.

configureLogsResponse_description :: Lens' ConfigureLogsResponse (Maybe Text)

A short text description of the Channel.

configureLogsResponse_egressAccessLogs :: Lens' ConfigureLogsResponse (Maybe EgressAccessLogs)

Undocumented member.

configureLogsResponse_tags :: Lens' ConfigureLogsResponse (Maybe (HashMap Text Text))

Undocumented member.

configureLogsResponse_httpStatus :: Lens' ConfigureLogsResponse Int

The response's http status code.

DescribeOriginEndpoint

describeOriginEndpoint_id :: Lens' DescribeOriginEndpoint Text

The ID of the OriginEndpoint.

describeOriginEndpointResponse_whitelist :: Lens' DescribeOriginEndpointResponse (Maybe [Text])

A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

describeOriginEndpointResponse_hlsPackage :: Lens' DescribeOriginEndpointResponse (Maybe HlsPackage)

Undocumented member.

describeOriginEndpointResponse_arn :: Lens' DescribeOriginEndpointResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the OriginEndpoint.

describeOriginEndpointResponse_manifestName :: Lens' DescribeOriginEndpointResponse (Maybe Text)

A short string appended to the end of the OriginEndpoint URL.

describeOriginEndpointResponse_url :: Lens' DescribeOriginEndpointResponse (Maybe Text)

The URL of the packaged OriginEndpoint for consumption.

describeOriginEndpointResponse_authorization :: Lens' DescribeOriginEndpointResponse (Maybe Authorization)

Undocumented member.

describeOriginEndpointResponse_channelId :: Lens' DescribeOriginEndpointResponse (Maybe Text)

The ID of the Channel the OriginEndpoint is associated with.

describeOriginEndpointResponse_startoverWindowSeconds :: Lens' DescribeOriginEndpointResponse (Maybe Int)

Maximum duration (seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

describeOriginEndpointResponse_dashPackage :: Lens' DescribeOriginEndpointResponse (Maybe DashPackage)

Undocumented member.

describeOriginEndpointResponse_mssPackage :: Lens' DescribeOriginEndpointResponse (Maybe MssPackage)

Undocumented member.

describeOriginEndpointResponse_id :: Lens' DescribeOriginEndpointResponse (Maybe Text)

The ID of the OriginEndpoint.

describeOriginEndpointResponse_timeDelaySeconds :: Lens' DescribeOriginEndpointResponse (Maybe Int)

Amount of delay (seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

describeOriginEndpointResponse_cmafPackage :: Lens' DescribeOriginEndpointResponse (Maybe CmafPackage)

Undocumented member.

describeOriginEndpointResponse_description :: Lens' DescribeOriginEndpointResponse (Maybe Text)

A short text description of the OriginEndpoint.

describeOriginEndpointResponse_tags :: Lens' DescribeOriginEndpointResponse (Maybe (HashMap Text Text))

Undocumented member.

describeOriginEndpointResponse_origination :: Lens' DescribeOriginEndpointResponse (Maybe Origination)

Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

describeOriginEndpointResponse_httpStatus :: Lens' DescribeOriginEndpointResponse Int

The response's http status code.

ListChannels

listChannels_nextToken :: Lens' ListChannels (Maybe Text)

A token used to resume pagination from the end of a previous request.

listChannels_maxResults :: Lens' ListChannels (Maybe Natural)

Upper bound on number of records to return.

listChannelsResponse_channels :: Lens' ListChannelsResponse (Maybe [Channel])

A list of Channel records.

listChannelsResponse_nextToken :: Lens' ListChannelsResponse (Maybe Text)

A token that can be used to resume pagination from the end of the collection.

listChannelsResponse_httpStatus :: Lens' ListChannelsResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

Undocumented member.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

Undocumented member.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

DeleteChannel

deleteChannel_id :: Lens' DeleteChannel Text

The ID of the Channel to delete.

deleteChannelResponse_httpStatus :: Lens' DeleteChannelResponse Int

The response's http status code.

UpdateChannel

updateChannel_description :: Lens' UpdateChannel (Maybe Text)

A short text description of the Channel.

updateChannel_id :: Lens' UpdateChannel Text

The ID of the Channel to update.

updateChannelResponse_ingressAccessLogs :: Lens' UpdateChannelResponse (Maybe IngressAccessLogs)

Undocumented member.

updateChannelResponse_hlsIngest :: Lens' UpdateChannelResponse (Maybe HlsIngest)

Undocumented member.

updateChannelResponse_arn :: Lens' UpdateChannelResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the Channel.

updateChannelResponse_id :: Lens' UpdateChannelResponse (Maybe Text)

The ID of the Channel.

updateChannelResponse_description :: Lens' UpdateChannelResponse (Maybe Text)

A short text description of the Channel.

updateChannelResponse_egressAccessLogs :: Lens' UpdateChannelResponse (Maybe EgressAccessLogs)

Undocumented member.

updateChannelResponse_tags :: Lens' UpdateChannelResponse (Maybe (HashMap Text Text))

Undocumented member.

updateChannelResponse_httpStatus :: Lens' UpdateChannelResponse Int

The response's http status code.

DescribeHarvestJob

describeHarvestJob_id :: Lens' DescribeHarvestJob Text

The ID of the HarvestJob.

describeHarvestJobResponse_status :: Lens' DescribeHarvestJobResponse (Maybe Status)

The current status of the HarvestJob. Consider setting up a CloudWatch Event to listen for HarvestJobs as they succeed or fail. In the event of failure, the CloudWatch Event will include an explanation of why the HarvestJob failed.

describeHarvestJobResponse_originEndpointId :: Lens' DescribeHarvestJobResponse (Maybe Text)

The ID of the OriginEndpoint that the HarvestJob will harvest from. This cannot be changed after the HarvestJob is submitted.

describeHarvestJobResponse_startTime :: Lens' DescribeHarvestJobResponse (Maybe Text)

The start of the time-window which will be harvested.

describeHarvestJobResponse_arn :: Lens' DescribeHarvestJobResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the HarvestJob.

describeHarvestJobResponse_createdAt :: Lens' DescribeHarvestJobResponse (Maybe Text)

The time the HarvestJob was submitted

describeHarvestJobResponse_channelId :: Lens' DescribeHarvestJobResponse (Maybe Text)

The ID of the Channel that the HarvestJob will harvest from.

describeHarvestJobResponse_s3Destination :: Lens' DescribeHarvestJobResponse (Maybe S3Destination)

Undocumented member.

describeHarvestJobResponse_endTime :: Lens' DescribeHarvestJobResponse (Maybe Text)

The end of the time-window which will be harvested.

describeHarvestJobResponse_id :: Lens' DescribeHarvestJobResponse (Maybe Text)

The ID of the HarvestJob. The ID must be unique within the region and it cannot be changed after the HarvestJob is submitted.

describeHarvestJobResponse_httpStatus :: Lens' DescribeHarvestJobResponse Int

The response's http status code.

RotateIngestEndpointCredentials

rotateIngestEndpointCredentials_ingestEndpointId :: Lens' RotateIngestEndpointCredentials Text

The id of the IngestEndpoint whose credentials should be rotated

rotateIngestEndpointCredentials_id :: Lens' RotateIngestEndpointCredentials Text

The ID of the channel the IngestEndpoint is on.

rotateIngestEndpointCredentialsResponse_ingressAccessLogs :: Lens' RotateIngestEndpointCredentialsResponse (Maybe IngressAccessLogs)

Undocumented member.

rotateIngestEndpointCredentialsResponse_hlsIngest :: Lens' RotateIngestEndpointCredentialsResponse (Maybe HlsIngest)

Undocumented member.

rotateIngestEndpointCredentialsResponse_arn :: Lens' RotateIngestEndpointCredentialsResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the Channel.

rotateIngestEndpointCredentialsResponse_id :: Lens' RotateIngestEndpointCredentialsResponse (Maybe Text)

The ID of the Channel.

rotateIngestEndpointCredentialsResponse_description :: Lens' RotateIngestEndpointCredentialsResponse (Maybe Text)

A short text description of the Channel.

rotateIngestEndpointCredentialsResponse_egressAccessLogs :: Lens' RotateIngestEndpointCredentialsResponse (Maybe EgressAccessLogs)

Undocumented member.

rotateIngestEndpointCredentialsResponse_tags :: Lens' RotateIngestEndpointCredentialsResponse (Maybe (HashMap Text Text))

Undocumented member.

rotateIngestEndpointCredentialsResponse_httpStatus :: Lens' RotateIngestEndpointCredentialsResponse Int

The response's http status code.

CreateOriginEndpoint

createOriginEndpoint_whitelist :: Lens' CreateOriginEndpoint (Maybe [Text])

A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

createOriginEndpoint_hlsPackage :: Lens' CreateOriginEndpoint (Maybe HlsPackage)

Undocumented member.

createOriginEndpoint_manifestName :: Lens' CreateOriginEndpoint (Maybe Text)

A short string that will be used as the filename of the OriginEndpoint URL (defaults to "index").

createOriginEndpoint_authorization :: Lens' CreateOriginEndpoint (Maybe Authorization)

Undocumented member.

createOriginEndpoint_startoverWindowSeconds :: Lens' CreateOriginEndpoint (Maybe Int)

Maximum duration (seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

createOriginEndpoint_dashPackage :: Lens' CreateOriginEndpoint (Maybe DashPackage)

Undocumented member.

createOriginEndpoint_mssPackage :: Lens' CreateOriginEndpoint (Maybe MssPackage)

Undocumented member.

createOriginEndpoint_timeDelaySeconds :: Lens' CreateOriginEndpoint (Maybe Int)

Amount of delay (seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

createOriginEndpoint_cmafPackage :: Lens' CreateOriginEndpoint (Maybe CmafPackageCreateOrUpdateParameters)

Undocumented member.

createOriginEndpoint_description :: Lens' CreateOriginEndpoint (Maybe Text)

A short text description of the OriginEndpoint.

createOriginEndpoint_tags :: Lens' CreateOriginEndpoint (Maybe (HashMap Text Text))

Undocumented member.

createOriginEndpoint_origination :: Lens' CreateOriginEndpoint (Maybe Origination)

Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

createOriginEndpoint_channelId :: Lens' CreateOriginEndpoint Text

The ID of the Channel that the OriginEndpoint will be associated with. This cannot be changed after the OriginEndpoint is created.

createOriginEndpoint_id :: Lens' CreateOriginEndpoint Text

The ID of the OriginEndpoint. The ID must be unique within the region and it cannot be changed after the OriginEndpoint is created.

createOriginEndpointResponse_whitelist :: Lens' CreateOriginEndpointResponse (Maybe [Text])

A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

createOriginEndpointResponse_hlsPackage :: Lens' CreateOriginEndpointResponse (Maybe HlsPackage)

Undocumented member.

createOriginEndpointResponse_arn :: Lens' CreateOriginEndpointResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the OriginEndpoint.

createOriginEndpointResponse_manifestName :: Lens' CreateOriginEndpointResponse (Maybe Text)

A short string appended to the end of the OriginEndpoint URL.

createOriginEndpointResponse_url :: Lens' CreateOriginEndpointResponse (Maybe Text)

The URL of the packaged OriginEndpoint for consumption.

createOriginEndpointResponse_authorization :: Lens' CreateOriginEndpointResponse (Maybe Authorization)

Undocumented member.

createOriginEndpointResponse_channelId :: Lens' CreateOriginEndpointResponse (Maybe Text)

The ID of the Channel the OriginEndpoint is associated with.

createOriginEndpointResponse_startoverWindowSeconds :: Lens' CreateOriginEndpointResponse (Maybe Int)

Maximum duration (seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

createOriginEndpointResponse_dashPackage :: Lens' CreateOriginEndpointResponse (Maybe DashPackage)

Undocumented member.

createOriginEndpointResponse_mssPackage :: Lens' CreateOriginEndpointResponse (Maybe MssPackage)

Undocumented member.

createOriginEndpointResponse_id :: Lens' CreateOriginEndpointResponse (Maybe Text)

The ID of the OriginEndpoint.

createOriginEndpointResponse_timeDelaySeconds :: Lens' CreateOriginEndpointResponse (Maybe Int)

Amount of delay (seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

createOriginEndpointResponse_cmafPackage :: Lens' CreateOriginEndpointResponse (Maybe CmafPackage)

Undocumented member.

createOriginEndpointResponse_description :: Lens' CreateOriginEndpointResponse (Maybe Text)

A short text description of the OriginEndpoint.

createOriginEndpointResponse_tags :: Lens' CreateOriginEndpointResponse (Maybe (HashMap Text Text))

Undocumented member.

createOriginEndpointResponse_origination :: Lens' CreateOriginEndpointResponse (Maybe Origination)

Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

createOriginEndpointResponse_httpStatus :: Lens' CreateOriginEndpointResponse Int

The response's http status code.

ListOriginEndpoints

listOriginEndpoints_channelId :: Lens' ListOriginEndpoints (Maybe Text)

When specified, the request will return only OriginEndpoints associated with the given Channel ID.

listOriginEndpoints_nextToken :: Lens' ListOriginEndpoints (Maybe Text)

A token used to resume pagination from the end of a previous request.

listOriginEndpoints_maxResults :: Lens' ListOriginEndpoints (Maybe Natural)

The upper bound on the number of records to return.

listOriginEndpointsResponse_originEndpoints :: Lens' ListOriginEndpointsResponse (Maybe [OriginEndpoint])

A list of OriginEndpoint records.

listOriginEndpointsResponse_nextToken :: Lens' ListOriginEndpointsResponse (Maybe Text)

A token that can be used to resume pagination from the end of the collection.

listOriginEndpointsResponse_httpStatus :: Lens' ListOriginEndpointsResponse Int

The response's http status code.

ListHarvestJobs

listHarvestJobs_includeStatus :: Lens' ListHarvestJobs (Maybe Text)

When specified, the request will return only HarvestJobs in the given status.

listHarvestJobs_nextToken :: Lens' ListHarvestJobs (Maybe Text)

A token used to resume pagination from the end of a previous request.

listHarvestJobs_includeChannelId :: Lens' ListHarvestJobs (Maybe Text)

When specified, the request will return only HarvestJobs associated with the given Channel ID.

listHarvestJobs_maxResults :: Lens' ListHarvestJobs (Maybe Natural)

The upper bound on the number of records to return.

listHarvestJobsResponse_harvestJobs :: Lens' ListHarvestJobsResponse (Maybe [HarvestJob])

A list of HarvestJob records.

listHarvestJobsResponse_nextToken :: Lens' ListHarvestJobsResponse (Maybe Text)

A token that can be used to resume pagination from the end of the collection.

listHarvestJobsResponse_httpStatus :: Lens' ListHarvestJobsResponse Int

The response's http status code.

CreateChannel

createChannel_description :: Lens' CreateChannel (Maybe Text)

A short text description of the Channel.

createChannel_tags :: Lens' CreateChannel (Maybe (HashMap Text Text))

Undocumented member.

createChannel_id :: Lens' CreateChannel Text

The ID of the Channel. The ID must be unique within the region and it cannot be changed after a Channel is created.

createChannelResponse_ingressAccessLogs :: Lens' CreateChannelResponse (Maybe IngressAccessLogs)

Undocumented member.

createChannelResponse_hlsIngest :: Lens' CreateChannelResponse (Maybe HlsIngest)

Undocumented member.

createChannelResponse_arn :: Lens' CreateChannelResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the Channel.

createChannelResponse_id :: Lens' CreateChannelResponse (Maybe Text)

The ID of the Channel.

createChannelResponse_description :: Lens' CreateChannelResponse (Maybe Text)

A short text description of the Channel.

createChannelResponse_egressAccessLogs :: Lens' CreateChannelResponse (Maybe EgressAccessLogs)

Undocumented member.

createChannelResponse_tags :: Lens' CreateChannelResponse (Maybe (HashMap Text Text))

Undocumented member.

createChannelResponse_httpStatus :: Lens' CreateChannelResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

Undocumented member.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

Undocumented member.

UntagResource

untagResource_tagKeys :: Lens' UntagResource [Text]

The key(s) of tag to be deleted

untagResource_resourceArn :: Lens' UntagResource Text

Undocumented member.

DescribeChannel

describeChannel_id :: Lens' DescribeChannel Text

The ID of a Channel.

describeChannelResponse_ingressAccessLogs :: Lens' DescribeChannelResponse (Maybe IngressAccessLogs)

Undocumented member.

describeChannelResponse_hlsIngest :: Lens' DescribeChannelResponse (Maybe HlsIngest)

Undocumented member.

describeChannelResponse_arn :: Lens' DescribeChannelResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the Channel.

describeChannelResponse_id :: Lens' DescribeChannelResponse (Maybe Text)

The ID of the Channel.

describeChannelResponse_description :: Lens' DescribeChannelResponse (Maybe Text)

A short text description of the Channel.

describeChannelResponse_egressAccessLogs :: Lens' DescribeChannelResponse (Maybe EgressAccessLogs)

Undocumented member.

describeChannelResponse_tags :: Lens' DescribeChannelResponse (Maybe (HashMap Text Text))

Undocumented member.

describeChannelResponse_httpStatus :: Lens' DescribeChannelResponse Int

The response's http status code.

DeleteOriginEndpoint

deleteOriginEndpoint_id :: Lens' DeleteOriginEndpoint Text

The ID of the OriginEndpoint to delete.

deleteOriginEndpointResponse_httpStatus :: Lens' DeleteOriginEndpointResponse Int

The response's http status code.

UpdateOriginEndpoint

updateOriginEndpoint_whitelist :: Lens' UpdateOriginEndpoint (Maybe [Text])

A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

updateOriginEndpoint_hlsPackage :: Lens' UpdateOriginEndpoint (Maybe HlsPackage)

Undocumented member.

updateOriginEndpoint_manifestName :: Lens' UpdateOriginEndpoint (Maybe Text)

A short string that will be appended to the end of the Endpoint URL.

updateOriginEndpoint_authorization :: Lens' UpdateOriginEndpoint (Maybe Authorization)

Undocumented member.

updateOriginEndpoint_startoverWindowSeconds :: Lens' UpdateOriginEndpoint (Maybe Int)

Maximum duration (in seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

updateOriginEndpoint_dashPackage :: Lens' UpdateOriginEndpoint (Maybe DashPackage)

Undocumented member.

updateOriginEndpoint_mssPackage :: Lens' UpdateOriginEndpoint (Maybe MssPackage)

Undocumented member.

updateOriginEndpoint_timeDelaySeconds :: Lens' UpdateOriginEndpoint (Maybe Int)

Amount of delay (in seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

updateOriginEndpoint_cmafPackage :: Lens' UpdateOriginEndpoint (Maybe CmafPackageCreateOrUpdateParameters)

Undocumented member.

updateOriginEndpoint_description :: Lens' UpdateOriginEndpoint (Maybe Text)

A short text description of the OriginEndpoint.

updateOriginEndpoint_origination :: Lens' UpdateOriginEndpoint (Maybe Origination)

Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

updateOriginEndpoint_id :: Lens' UpdateOriginEndpoint Text

The ID of the OriginEndpoint to update.

updateOriginEndpointResponse_whitelist :: Lens' UpdateOriginEndpointResponse (Maybe [Text])

A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

updateOriginEndpointResponse_hlsPackage :: Lens' UpdateOriginEndpointResponse (Maybe HlsPackage)

Undocumented member.

updateOriginEndpointResponse_arn :: Lens' UpdateOriginEndpointResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the OriginEndpoint.

updateOriginEndpointResponse_manifestName :: Lens' UpdateOriginEndpointResponse (Maybe Text)

A short string appended to the end of the OriginEndpoint URL.

updateOriginEndpointResponse_url :: Lens' UpdateOriginEndpointResponse (Maybe Text)

The URL of the packaged OriginEndpoint for consumption.

updateOriginEndpointResponse_authorization :: Lens' UpdateOriginEndpointResponse (Maybe Authorization)

Undocumented member.

updateOriginEndpointResponse_channelId :: Lens' UpdateOriginEndpointResponse (Maybe Text)

The ID of the Channel the OriginEndpoint is associated with.

updateOriginEndpointResponse_startoverWindowSeconds :: Lens' UpdateOriginEndpointResponse (Maybe Int)

Maximum duration (seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

updateOriginEndpointResponse_dashPackage :: Lens' UpdateOriginEndpointResponse (Maybe DashPackage)

Undocumented member.

updateOriginEndpointResponse_mssPackage :: Lens' UpdateOriginEndpointResponse (Maybe MssPackage)

Undocumented member.

updateOriginEndpointResponse_id :: Lens' UpdateOriginEndpointResponse (Maybe Text)

The ID of the OriginEndpoint.

updateOriginEndpointResponse_timeDelaySeconds :: Lens' UpdateOriginEndpointResponse (Maybe Int)

Amount of delay (seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

updateOriginEndpointResponse_cmafPackage :: Lens' UpdateOriginEndpointResponse (Maybe CmafPackage)

Undocumented member.

updateOriginEndpointResponse_description :: Lens' UpdateOriginEndpointResponse (Maybe Text)

A short text description of the OriginEndpoint.

updateOriginEndpointResponse_tags :: Lens' UpdateOriginEndpointResponse (Maybe (HashMap Text Text))

Undocumented member.

updateOriginEndpointResponse_origination :: Lens' UpdateOriginEndpointResponse (Maybe Origination)

Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

updateOriginEndpointResponse_httpStatus :: Lens' UpdateOriginEndpointResponse Int

The response's http status code.

Types

Authorization

authorization_secretsRoleArn :: Lens' Authorization Text

The Amazon Resource Name (ARN) for the IAM role that allows MediaPackage to communicate with AWS Secrets Manager.

authorization_cdnIdentifierSecret :: Lens' Authorization Text

The Amazon Resource Name (ARN) for the secret in Secrets Manager that your Content Distribution Network (CDN) uses for authorization to access your endpoint.

Channel

channel_ingressAccessLogs :: Lens' Channel (Maybe IngressAccessLogs)

Undocumented member.

channel_hlsIngest :: Lens' Channel (Maybe HlsIngest)

Undocumented member.

channel_arn :: Lens' Channel (Maybe Text)

The Amazon Resource Name (ARN) assigned to the Channel.

channel_id :: Lens' Channel (Maybe Text)

The ID of the Channel.

channel_description :: Lens' Channel (Maybe Text)

A short text description of the Channel.

channel_egressAccessLogs :: Lens' Channel (Maybe EgressAccessLogs)

Undocumented member.

channel_tags :: Lens' Channel (Maybe (HashMap Text Text))

Undocumented member.

CmafEncryption

cmafEncryption_keyRotationIntervalSeconds :: Lens' CmafEncryption (Maybe Int)

Time (in seconds) between each encryption key rotation.

cmafEncryption_constantInitializationVector :: Lens' CmafEncryption (Maybe Text)

An optional 128-bit, 16-byte hex value represented by a 32-character string, used in conjunction with the key for encrypting blocks. If you don't specify a value, then MediaPackage creates the constant initialization vector (IV).

cmafEncryption_spekeKeyProvider :: Lens' CmafEncryption SpekeKeyProvider

Undocumented member.

CmafPackage

cmafPackage_hlsManifests :: Lens' CmafPackage (Maybe [HlsManifest])

A list of HLS manifest configurations

cmafPackage_segmentDurationSeconds :: Lens' CmafPackage (Maybe Int)

Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source segment duration.

cmafPackage_streamSelection :: Lens' CmafPackage (Maybe StreamSelection)

Undocumented member.

cmafPackage_encryption :: Lens' CmafPackage (Maybe CmafEncryption)

Undocumented member.

cmafPackage_segmentPrefix :: Lens' CmafPackage (Maybe Text)

An optional custom string that is prepended to the name of each segment. If not specified, it defaults to the ChannelId.

CmafPackageCreateOrUpdateParameters

cmafPackageCreateOrUpdateParameters_hlsManifests :: Lens' CmafPackageCreateOrUpdateParameters (Maybe [HlsManifestCreateOrUpdateParameters])

A list of HLS manifest configurations

cmafPackageCreateOrUpdateParameters_segmentDurationSeconds :: Lens' CmafPackageCreateOrUpdateParameters (Maybe Int)

Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source segment duration.

cmafPackageCreateOrUpdateParameters_streamSelection :: Lens' CmafPackageCreateOrUpdateParameters (Maybe StreamSelection)

Undocumented member.

cmafPackageCreateOrUpdateParameters_encryption :: Lens' CmafPackageCreateOrUpdateParameters (Maybe CmafEncryption)

Undocumented member.

cmafPackageCreateOrUpdateParameters_segmentPrefix :: Lens' CmafPackageCreateOrUpdateParameters (Maybe Text)

An optional custom string that is prepended to the name of each segment. If not specified, it defaults to the ChannelId.

DashEncryption

dashEncryption_keyRotationIntervalSeconds :: Lens' DashEncryption (Maybe Int)

Time (in seconds) between each encryption key rotation.

dashEncryption_spekeKeyProvider :: Lens' DashEncryption SpekeKeyProvider

Undocumented member.

DashPackage

dashPackage_adsOnDeliveryRestrictions :: Lens' DashPackage (Maybe AdsOnDeliveryRestrictions)

Undocumented member.

dashPackage_minBufferTimeSeconds :: Lens' DashPackage (Maybe Int)

Minimum duration (in seconds) that a player will buffer media before starting the presentation.

dashPackage_utcTiming :: Lens' DashPackage (Maybe UtcTiming)

Determines the type of UTCTiming included in the Media Presentation Description (MPD)

dashPackage_segmentTemplateFormat :: Lens' DashPackage (Maybe SegmentTemplateFormat)

Determines the type of SegmentTemplate included in the Media Presentation Description (MPD). When set to NUMBER_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Number$ media URLs. When set to TIME_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Time$ media URLs. When set to NUMBER_WITH_DURATION, only a duration is included in each SegmentTemplate, with $Number$ media URLs.

dashPackage_profile :: Lens' DashPackage (Maybe Profile)

The Dynamic Adaptive Streaming over HTTP (DASH) profile type. When set to "HBBTV_1_5", HbbTV 1.5 compliant output is enabled.

dashPackage_segmentDurationSeconds :: Lens' DashPackage (Maybe Int)

Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source segment duration.

dashPackage_utcTimingUri :: Lens' DashPackage (Maybe Text)

Specifies the value attribute of the UTCTiming field when utcTiming is set to HTTP-ISO or HTTP-HEAD

dashPackage_streamSelection :: Lens' DashPackage (Maybe StreamSelection)

Undocumented member.

dashPackage_encryption :: Lens' DashPackage (Maybe DashEncryption)

Undocumented member.

dashPackage_minUpdatePeriodSeconds :: Lens' DashPackage (Maybe Int)

Minimum duration (in seconds) between potential changes to the Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation Description (MPD).

dashPackage_manifestLayout :: Lens' DashPackage (Maybe ManifestLayout)

Determines the position of some tags in the Media Presentation Description (MPD). When set to FULL, elements like SegmentTemplate and ContentProtection are included in each Representation. When set to COMPACT, duplicate elements are combined and presented at the AdaptationSet level.

dashPackage_suggestedPresentationDelaySeconds :: Lens' DashPackage (Maybe Int)

Duration (in seconds) to delay live content before presentation.

dashPackage_manifestWindowSeconds :: Lens' DashPackage (Maybe Int)

Time window (in seconds) contained in each manifest.

dashPackage_adTriggers :: Lens' DashPackage (Maybe [AdTriggersElement])

Undocumented member.

dashPackage_periodTriggers :: Lens' DashPackage (Maybe [PeriodTriggersElement])

A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation Description (MPD) will be partitioned into multiple periods. If empty, the content will not be partitioned into more than one period. If the list contains "ADS", new periods will be created where the Channel source contains SCTE-35 ad markers.

EgressAccessLogs

egressAccessLogs_logGroupName :: Lens' EgressAccessLogs (Maybe Text)

Customize the log group name.

EncryptionContractConfiguration

encryptionContractConfiguration_presetSpeke20Audio :: Lens' EncryptionContractConfiguration PresetSpeke20Audio

A collection of audio encryption presets.

encryptionContractConfiguration_presetSpeke20Video :: Lens' EncryptionContractConfiguration PresetSpeke20Video

A collection of video encryption presets.

HarvestJob

harvestJob_status :: Lens' HarvestJob (Maybe Status)

The current status of the HarvestJob. Consider setting up a CloudWatch Event to listen for HarvestJobs as they succeed or fail. In the event of failure, the CloudWatch Event will include an explanation of why the HarvestJob failed.

harvestJob_originEndpointId :: Lens' HarvestJob (Maybe Text)

The ID of the OriginEndpoint that the HarvestJob will harvest from. This cannot be changed after the HarvestJob is submitted.

harvestJob_startTime :: Lens' HarvestJob (Maybe Text)

The start of the time-window which will be harvested.

harvestJob_arn :: Lens' HarvestJob (Maybe Text)

The Amazon Resource Name (ARN) assigned to the HarvestJob.

harvestJob_createdAt :: Lens' HarvestJob (Maybe Text)

The time the HarvestJob was submitted

harvestJob_channelId :: Lens' HarvestJob (Maybe Text)

The ID of the Channel that the HarvestJob will harvest from.

harvestJob_s3Destination :: Lens' HarvestJob (Maybe S3Destination)

Undocumented member.

harvestJob_endTime :: Lens' HarvestJob (Maybe Text)

The end of the time-window which will be harvested.

harvestJob_id :: Lens' HarvestJob (Maybe Text)

The ID of the HarvestJob. The ID must be unique within the region and it cannot be changed after the HarvestJob is submitted.

HlsEncryption

hlsEncryption_encryptionMethod :: Lens' HlsEncryption (Maybe EncryptionMethod)

The encryption method to use.

hlsEncryption_keyRotationIntervalSeconds :: Lens' HlsEncryption (Maybe Int)

Interval (in seconds) between each encryption key rotation.

hlsEncryption_constantInitializationVector :: Lens' HlsEncryption (Maybe Text)

A constant initialization vector for encryption (optional). When not specified the initialization vector will be periodically rotated.

hlsEncryption_repeatExtXKey :: Lens' HlsEncryption (Maybe Bool)

When enabled, the EXT-X-KEY tag will be repeated in output manifests.

hlsEncryption_spekeKeyProvider :: Lens' HlsEncryption SpekeKeyProvider

Undocumented member.

HlsIngest

hlsIngest_ingestEndpoints :: Lens' HlsIngest (Maybe [IngestEndpoint])

A list of endpoints to which the source stream should be sent.

HlsManifest

hlsManifest_manifestName :: Lens' HlsManifest (Maybe Text)

An optional short string appended to the end of the OriginEndpoint URL. If not specified, defaults to the manifestName for the OriginEndpoint.

hlsManifest_url :: Lens' HlsManifest (Maybe Text)

The URL of the packaged OriginEndpoint for consumption.

hlsManifest_playlistType :: Lens' HlsManifest (Maybe PlaylistType)

The HTTP Live Streaming (HLS) playlist type. When either "EVENT" or "VOD" is specified, a corresponding EXT-X-PLAYLIST-TYPE entry will be included in the media playlist.

hlsManifest_programDateTimeIntervalSeconds :: Lens' HlsManifest (Maybe Int)

The interval (in seconds) between each EXT-X-PROGRAM-DATE-TIME tag inserted into manifests. Additionally, when an interval is specified ID3Timed Metadata messages will be generated every 5 seconds using the ingest time of the content. If the interval is not specified, or set to 0, then no EXT-X-PROGRAM-DATE-TIME tags will be inserted into manifests and no ID3Timed Metadata messages will be generated. Note that irrespective of this parameter, if any ID3 Timed Metadata is found in HTTP Live Streaming (HLS) input, it will be passed through to HLS output.

hlsManifest_adMarkers :: Lens' HlsManifest (Maybe AdMarkers)

This setting controls how ad markers are included in the packaged OriginEndpoint. "NONE" will omit all SCTE-35 ad markers from the output. "PASSTHROUGH" causes the manifest to contain a copy of the SCTE-35 ad markers (comments) taken directly from the input HTTP Live Streaming (HLS) manifest. "SCTE35_ENHANCED" generates ad markers and blackout tags based on SCTE-35 messages in the input source. "DATERANGE" inserts EXT-X-DATERANGE tags to signal ad and program transition events in HLS and CMAF manifests. For this option, you must set a programDateTimeIntervalSeconds value that is greater than 0.

hlsManifest_includeIframeOnlyStream :: Lens' HlsManifest (Maybe Bool)

When enabled, an I-Frame only stream will be included in the output.

hlsManifest_playlistWindowSeconds :: Lens' HlsManifest (Maybe Int)

Time window (in seconds) contained in each parent manifest.

hlsManifest_id :: Lens' HlsManifest Text

The ID of the manifest. The ID must be unique within the OriginEndpoint and it cannot be changed after it is created.

HlsManifestCreateOrUpdateParameters

hlsManifestCreateOrUpdateParameters_adsOnDeliveryRestrictions :: Lens' HlsManifestCreateOrUpdateParameters (Maybe AdsOnDeliveryRestrictions)

Undocumented member.

hlsManifestCreateOrUpdateParameters_manifestName :: Lens' HlsManifestCreateOrUpdateParameters (Maybe Text)

An optional short string appended to the end of the OriginEndpoint URL. If not specified, defaults to the manifestName for the OriginEndpoint.

hlsManifestCreateOrUpdateParameters_playlistType :: Lens' HlsManifestCreateOrUpdateParameters (Maybe PlaylistType)

The HTTP Live Streaming (HLS) playlist type. When either "EVENT" or "VOD" is specified, a corresponding EXT-X-PLAYLIST-TYPE entry will be included in the media playlist.

hlsManifestCreateOrUpdateParameters_programDateTimeIntervalSeconds :: Lens' HlsManifestCreateOrUpdateParameters (Maybe Int)

The interval (in seconds) between each EXT-X-PROGRAM-DATE-TIME tag inserted into manifests. Additionally, when an interval is specified ID3Timed Metadata messages will be generated every 5 seconds using the ingest time of the content. If the interval is not specified, or set to 0, then no EXT-X-PROGRAM-DATE-TIME tags will be inserted into manifests and no ID3Timed Metadata messages will be generated. Note that irrespective of this parameter, if any ID3 Timed Metadata is found in HTTP Live Streaming (HLS) input, it will be passed through to HLS output.

hlsManifestCreateOrUpdateParameters_adMarkers :: Lens' HlsManifestCreateOrUpdateParameters (Maybe AdMarkers)

This setting controls how ad markers are included in the packaged OriginEndpoint. "NONE" will omit all SCTE-35 ad markers from the output. "PASSTHROUGH" causes the manifest to contain a copy of the SCTE-35 ad markers (comments) taken directly from the input HTTP Live Streaming (HLS) manifest. "SCTE35_ENHANCED" generates ad markers and blackout tags based on SCTE-35 messages in the input source. "DATERANGE" inserts EXT-X-DATERANGE tags to signal ad and program transition events in HLS and CMAF manifests. For this option, you must set a programDateTimeIntervalSeconds value that is greater than 0.

hlsManifestCreateOrUpdateParameters_includeIframeOnlyStream :: Lens' HlsManifestCreateOrUpdateParameters (Maybe Bool)

When enabled, an I-Frame only stream will be included in the output.

hlsManifestCreateOrUpdateParameters_adTriggers :: Lens' HlsManifestCreateOrUpdateParameters (Maybe [AdTriggersElement])

Undocumented member.

hlsManifestCreateOrUpdateParameters_playlistWindowSeconds :: Lens' HlsManifestCreateOrUpdateParameters (Maybe Int)

Time window (in seconds) contained in each parent manifest.

hlsManifestCreateOrUpdateParameters_id :: Lens' HlsManifestCreateOrUpdateParameters Text

The ID of the manifest. The ID must be unique within the OriginEndpoint and it cannot be changed after it is created.

HlsPackage

hlsPackage_adsOnDeliveryRestrictions :: Lens' HlsPackage (Maybe AdsOnDeliveryRestrictions)

Undocumented member.

hlsPackage_useAudioRenditionGroup :: Lens' HlsPackage (Maybe Bool)

When enabled, audio streams will be placed in rendition groups in the output.

hlsPackage_includeDvbSubtitles :: Lens' HlsPackage (Maybe Bool)

When enabled, MediaPackage passes through digital video broadcasting (DVB) subtitles into the output.

hlsPackage_playlistType :: Lens' HlsPackage (Maybe PlaylistType)

The HTTP Live Streaming (HLS) playlist type. When either "EVENT" or "VOD" is specified, a corresponding EXT-X-PLAYLIST-TYPE entry will be included in the media playlist.

hlsPackage_segmentDurationSeconds :: Lens' HlsPackage (Maybe Int)

Duration (in seconds) of each fragment. Actual fragments will be rounded to the nearest multiple of the source fragment duration.

hlsPackage_programDateTimeIntervalSeconds :: Lens' HlsPackage (Maybe Int)

The interval (in seconds) between each EXT-X-PROGRAM-DATE-TIME tag inserted into manifests. Additionally, when an interval is specified ID3Timed Metadata messages will be generated every 5 seconds using the ingest time of the content. If the interval is not specified, or set to 0, then no EXT-X-PROGRAM-DATE-TIME tags will be inserted into manifests and no ID3Timed Metadata messages will be generated. Note that irrespective of this parameter, if any ID3 Timed Metadata is found in HTTP Live Streaming (HLS) input, it will be passed through to HLS output.

hlsPackage_streamSelection :: Lens' HlsPackage (Maybe StreamSelection)

Undocumented member.

hlsPackage_adMarkers :: Lens' HlsPackage (Maybe AdMarkers)

This setting controls how ad markers are included in the packaged OriginEndpoint. "NONE" will omit all SCTE-35 ad markers from the output. "PASSTHROUGH" causes the manifest to contain a copy of the SCTE-35 ad markers (comments) taken directly from the input HTTP Live Streaming (HLS) manifest. "SCTE35_ENHANCED" generates ad markers and blackout tags based on SCTE-35 messages in the input source. "DATERANGE" inserts EXT-X-DATERANGE tags to signal ad and program transition events in HLS and CMAF manifests. For this option, you must set a programDateTimeIntervalSeconds value that is greater than 0.

hlsPackage_encryption :: Lens' HlsPackage (Maybe HlsEncryption)

Undocumented member.

hlsPackage_includeIframeOnlyStream :: Lens' HlsPackage (Maybe Bool)

When enabled, an I-Frame only stream will be included in the output.

hlsPackage_adTriggers :: Lens' HlsPackage (Maybe [AdTriggersElement])

Undocumented member.

hlsPackage_playlistWindowSeconds :: Lens' HlsPackage (Maybe Int)

Time window (in seconds) contained in each parent manifest.

IngestEndpoint

ingestEndpoint_url :: Lens' IngestEndpoint (Maybe Text)

The ingest URL to which the source stream should be sent.

ingestEndpoint_username :: Lens' IngestEndpoint (Maybe Text)

The system generated username for ingest authentication.

ingestEndpoint_password :: Lens' IngestEndpoint (Maybe Text)

The system generated password for ingest authentication.

ingestEndpoint_id :: Lens' IngestEndpoint (Maybe Text)

The system generated unique identifier for the IngestEndpoint

IngressAccessLogs

ingressAccessLogs_logGroupName :: Lens' IngressAccessLogs (Maybe Text)

Customize the log group name.

MssEncryption

mssEncryption_spekeKeyProvider :: Lens' MssEncryption SpekeKeyProvider

Undocumented member.

MssPackage

mssPackage_segmentDurationSeconds :: Lens' MssPackage (Maybe Int)

The duration (in seconds) of each segment.

mssPackage_streamSelection :: Lens' MssPackage (Maybe StreamSelection)

Undocumented member.

mssPackage_encryption :: Lens' MssPackage (Maybe MssEncryption)

Undocumented member.

mssPackage_manifestWindowSeconds :: Lens' MssPackage (Maybe Int)

The time window (in seconds) contained in each manifest.

OriginEndpoint

originEndpoint_whitelist :: Lens' OriginEndpoint (Maybe [Text])

A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

originEndpoint_hlsPackage :: Lens' OriginEndpoint (Maybe HlsPackage)

Undocumented member.

originEndpoint_arn :: Lens' OriginEndpoint (Maybe Text)

The Amazon Resource Name (ARN) assigned to the OriginEndpoint.

originEndpoint_manifestName :: Lens' OriginEndpoint (Maybe Text)

A short string appended to the end of the OriginEndpoint URL.

originEndpoint_url :: Lens' OriginEndpoint (Maybe Text)

The URL of the packaged OriginEndpoint for consumption.

originEndpoint_authorization :: Lens' OriginEndpoint (Maybe Authorization)

Undocumented member.

originEndpoint_channelId :: Lens' OriginEndpoint (Maybe Text)

The ID of the Channel the OriginEndpoint is associated with.

originEndpoint_startoverWindowSeconds :: Lens' OriginEndpoint (Maybe Int)

Maximum duration (seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

originEndpoint_dashPackage :: Lens' OriginEndpoint (Maybe DashPackage)

Undocumented member.

originEndpoint_mssPackage :: Lens' OriginEndpoint (Maybe MssPackage)

Undocumented member.

originEndpoint_id :: Lens' OriginEndpoint (Maybe Text)

The ID of the OriginEndpoint.

originEndpoint_timeDelaySeconds :: Lens' OriginEndpoint (Maybe Int)

Amount of delay (seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

originEndpoint_cmafPackage :: Lens' OriginEndpoint (Maybe CmafPackage)

Undocumented member.

originEndpoint_description :: Lens' OriginEndpoint (Maybe Text)

A short text description of the OriginEndpoint.

originEndpoint_tags :: Lens' OriginEndpoint (Maybe (HashMap Text Text))

Undocumented member.

originEndpoint_origination :: Lens' OriginEndpoint (Maybe Origination)

Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

S3Destination

s3Destination_manifestKey :: Lens' S3Destination Text

The key in the specified S3 bucket where the harvested top-level manifest will be placed.

s3Destination_bucketName :: Lens' S3Destination Text

The name of an S3 bucket within which harvested content will be exported

s3Destination_roleArn :: Lens' S3Destination Text

The IAM role used to write to the specified S3 bucket

SpekeKeyProvider

spekeKeyProvider_encryptionContractConfiguration :: Lens' SpekeKeyProvider (Maybe EncryptionContractConfiguration)

Undocumented member.

spekeKeyProvider_certificateArn :: Lens' SpekeKeyProvider (Maybe Text)

An Amazon Resource Name (ARN) of a Certificate Manager certificate that MediaPackage will use for enforcing secure end-to-end data transfer with the key provider service.

spekeKeyProvider_resourceId :: Lens' SpekeKeyProvider Text

The resource ID to include in key requests.

spekeKeyProvider_systemIds :: Lens' SpekeKeyProvider [Text]

The system IDs to include in key requests.

spekeKeyProvider_url :: Lens' SpekeKeyProvider Text

The URL of the external key provider service.

spekeKeyProvider_roleArn :: Lens' SpekeKeyProvider Text

An Amazon Resource Name (ARN) of an IAM role that AWS Elemental MediaPackage will assume when accessing the key provider service.

StreamSelection

streamSelection_streamOrder :: Lens' StreamSelection (Maybe StreamOrder)

A directive that determines the order of streams in the output.

streamSelection_minVideoBitsPerSecond :: Lens' StreamSelection (Maybe Int)

The minimum video bitrate (bps) to include in output.

streamSelection_maxVideoBitsPerSecond :: Lens' StreamSelection (Maybe Int)

The maximum video bitrate (bps) to include in output.