Network.AWS.MediaPackage.ListHarvestJobs

Description

Returns a collection of HarvestJob records.

This operation returns paginated results.

Creating a Request

data ListHarvestJobs

See: newListHarvestJobs smart constructor.

Constructors

ListHarvestJobs'

Fields

nextToken :: Maybe Text

A token used to resume pagination from the end of a previous request.

maxResults :: Maybe Natural

The upper bound on the number of records to return.

includeStatus :: Maybe Text

When specified, the request will return only HarvestJobs in the given status.

includeChannelId :: Maybe Text

When specified, the request will return only HarvestJobs associated with the given Channel ID.

newListHarvestJobs :: ListHarvestJobs

Create a value of ListHarvestJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListHarvestJobs', listHarvestJobs_nextToken - A token used to resume pagination from the end of a previous request.

$sel:maxResults:ListHarvestJobs', listHarvestJobs_maxResults - The upper bound on the number of records to return.

$sel:includeStatus:ListHarvestJobs', listHarvestJobs_includeStatus - When specified, the request will return only HarvestJobs in the given status.

$sel:includeChannelId:ListHarvestJobs', listHarvestJobs_includeChannelId - When specified, the request will return only HarvestJobs associated with the given Channel ID.

Request Lenses

listHarvestJobs_nextToken :: Lens' ListHarvestJobs (Maybe Text)

A token used to resume pagination from the end of a previous request.

listHarvestJobs_maxResults :: Lens' ListHarvestJobs (Maybe Natural)

The upper bound on the number of records to return.

listHarvestJobs_includeStatus :: Lens' ListHarvestJobs (Maybe Text)

When specified, the request will return only HarvestJobs in the given status.

listHarvestJobs_includeChannelId :: Lens' ListHarvestJobs (Maybe Text)

When specified, the request will return only HarvestJobs associated with the given Channel ID.

Destructuring the Response

data ListHarvestJobsResponse

See: newListHarvestJobsResponse smart constructor.

Constructors

ListHarvestJobsResponse'

Fields

nextToken :: Maybe Text

A token that can be used to resume pagination from the end of the collection.

harvestJobs :: Maybe [HarvestJob]

A list of HarvestJob records.

httpStatus :: Int

The response's http status code.

newListHarvestJobsResponse

Arguments

:: Int
$sel:httpStatus:ListHarvestJobsResponse'

-> ListHarvestJobsResponse

Create a value of ListHarvestJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListHarvestJobs', listHarvestJobsResponse_nextToken - A token that can be used to resume pagination from the end of the collection.

$sel:harvestJobs:ListHarvestJobsResponse', listHarvestJobsResponse_harvestJobs - A list of HarvestJob records.

$sel:httpStatus:ListHarvestJobsResponse', listHarvestJobsResponse_httpStatus - The response's http status code.

Response Lenses

listHarvestJobsResponse_nextToken :: Lens' ListHarvestJobsResponse (Maybe Text)

A token that can be used to resume pagination from the end of the collection.

listHarvestJobsResponse_harvestJobs :: Lens' ListHarvestJobsResponse (Maybe [HarvestJob])

A list of HarvestJob records.

listHarvestJobsResponse_httpStatus :: Lens' ListHarvestJobsResponse Int

The response's http status code.