Network.AWS.MediaPackage.Types.Channel

Description

Documentation

data Channel

A Channel resource configuration.

See: newChannel smart constructor.

Constructors

Channel'

Fields

egressAccessLogs :: Maybe EgressAccessLogs

arn :: Maybe Text

The Amazon Resource Name (ARN) assigned to the Channel.

id :: Maybe Text

The ID of the Channel.

hlsIngest :: Maybe HlsIngest

ingressAccessLogs :: Maybe IngressAccessLogs

tags :: Maybe (HashMap Text Text)

description :: Maybe Text

A short text description of the Channel.

newChannel :: Channel

Create a value of Channel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:egressAccessLogs:Channel', channel_egressAccessLogs - Undocumented member.

$sel:arn:Channel', channel_arn - The Amazon Resource Name (ARN) assigned to the Channel.

$sel:id:Channel', channel_id - The ID of the Channel.

$sel:hlsIngest:Channel', channel_hlsIngest - Undocumented member.

$sel:ingressAccessLogs:Channel', channel_ingressAccessLogs - Undocumented member.

$sel:tags:Channel', channel_tags - Undocumented member.

$sel:description:Channel', channel_description - A short text description of the Channel.

channel_egressAccessLogs :: Lens' Channel (Maybe EgressAccessLogs)

Undocumented member.

channel_arn :: Lens' Channel (Maybe Text)

The Amazon Resource Name (ARN) assigned to the Channel.

channel_id :: Lens' Channel (Maybe Text)

The ID of the Channel.

channel_hlsIngest :: Lens' Channel (Maybe HlsIngest)

Undocumented member.

channel_ingressAccessLogs :: Lens' Channel (Maybe IngressAccessLogs)

Undocumented member.

channel_tags :: Lens' Channel (Maybe (HashMap Text Text))

Undocumented member.

channel_description :: Lens' Channel (Maybe Text)

A short text description of the Channel.