Network.AWS.MediaPackage.Types.CmafPackage

Description

Documentation

data CmafPackage

A Common Media Application Format (CMAF) packaging configuration.

See: newCmafPackage smart constructor.

Constructors

CmafPackage'

Fields

hlsManifests :: Maybe [HlsManifest]

A list of HLS manifest configurations

segmentDurationSeconds :: Maybe Int

Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source segment duration.

streamSelection :: Maybe StreamSelection

encryption :: Maybe CmafEncryption

segmentPrefix :: Maybe Text

An optional custom string that is prepended to the name of each segment. If not specified, it defaults to the ChannelId.

newCmafPackage :: CmafPackage

Create a value of CmafPackage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hlsManifests:CmafPackage', cmafPackage_hlsManifests - A list of HLS manifest configurations

$sel:segmentDurationSeconds:CmafPackage', cmafPackage_segmentDurationSeconds - Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source segment duration.

$sel:streamSelection:CmafPackage', cmafPackage_streamSelection - Undocumented member.

$sel:encryption:CmafPackage', cmafPackage_encryption - Undocumented member.

$sel:segmentPrefix:CmafPackage', cmafPackage_segmentPrefix - An optional custom string that is prepended to the name of each segment. If not specified, it defaults to the ChannelId.

cmafPackage_hlsManifests :: Lens' CmafPackage (Maybe [HlsManifest])

A list of HLS manifest configurations

cmafPackage_segmentDurationSeconds :: Lens' CmafPackage (Maybe Int)

Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source segment duration.

cmafPackage_streamSelection :: Lens' CmafPackage (Maybe StreamSelection)

Undocumented member.

cmafPackage_encryption :: Lens' CmafPackage (Maybe CmafEncryption)

Undocumented member.

cmafPackage_segmentPrefix :: Lens' CmafPackage (Maybe Text)

An optional custom string that is prepended to the name of each segment. If not specified, it defaults to the ChannelId.