Network.AWS.MediaPackage.Types.CmafPackageCreateOrUpdateParameters

Description

Documentation

data CmafPackageCreateOrUpdateParameters

A Common Media Application Format (CMAF) packaging configuration.

See: newCmafPackageCreateOrUpdateParameters smart constructor.

Constructors

CmafPackageCreateOrUpdateParameters'

Fields

hlsManifests :: Maybe [HlsManifestCreateOrUpdateParameters]

A list of HLS manifest configurations

segmentDurationSeconds :: Maybe Int

Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source segment duration.

streamSelection :: Maybe StreamSelection

encryption :: Maybe CmafEncryption

segmentPrefix :: Maybe Text

An optional custom string that is prepended to the name of each segment. If not specified, it defaults to the ChannelId.

newCmafPackageCreateOrUpdateParameters :: CmafPackageCreateOrUpdateParameters

Create a value of CmafPackageCreateOrUpdateParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hlsManifests:CmafPackageCreateOrUpdateParameters', cmafPackageCreateOrUpdateParameters_hlsManifests - A list of HLS manifest configurations

$sel:segmentDurationSeconds:CmafPackageCreateOrUpdateParameters', cmafPackageCreateOrUpdateParameters_segmentDurationSeconds - Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source segment duration.

$sel:streamSelection:CmafPackageCreateOrUpdateParameters', cmafPackageCreateOrUpdateParameters_streamSelection - Undocumented member.

$sel:encryption:CmafPackageCreateOrUpdateParameters', cmafPackageCreateOrUpdateParameters_encryption - Undocumented member.

$sel:segmentPrefix:CmafPackageCreateOrUpdateParameters', cmafPackageCreateOrUpdateParameters_segmentPrefix - An optional custom string that is prepended to the name of each segment. If not specified, it defaults to the ChannelId.

cmafPackageCreateOrUpdateParameters_hlsManifests :: Lens' CmafPackageCreateOrUpdateParameters (Maybe [HlsManifestCreateOrUpdateParameters])

A list of HLS manifest configurations

cmafPackageCreateOrUpdateParameters_segmentDurationSeconds :: Lens' CmafPackageCreateOrUpdateParameters (Maybe Int)

Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source segment duration.

cmafPackageCreateOrUpdateParameters_streamSelection :: Lens' CmafPackageCreateOrUpdateParameters (Maybe StreamSelection)

Undocumented member.

cmafPackageCreateOrUpdateParameters_encryption :: Lens' CmafPackageCreateOrUpdateParameters (Maybe CmafEncryption)

Undocumented member.

cmafPackageCreateOrUpdateParameters_segmentPrefix :: Lens' CmafPackageCreateOrUpdateParameters (Maybe Text)

An optional custom string that is prepended to the name of each segment. If not specified, it defaults to the ChannelId.