Network.AWS.MediaPackage.Types.HarvestJob

Description

Documentation

data HarvestJob

A HarvestJob resource configuration

See: newHarvestJob smart constructor.

Constructors

HarvestJob'

Fields

s3Destination :: Maybe S3Destination

status :: Maybe Status

The current status of the HarvestJob. Consider setting up a CloudWatch Event to listen for HarvestJobs as they succeed or fail. In the event of failure, the CloudWatch Event will include an explanation of why the HarvestJob failed.

channelId :: Maybe Text

The ID of the Channel that the HarvestJob will harvest from.

createdAt :: Maybe Text

The time the HarvestJob was submitted

arn :: Maybe Text

The Amazon Resource Name (ARN) assigned to the HarvestJob.

id :: Maybe Text

The ID of the HarvestJob. The ID must be unique within the region and it cannot be changed after the HarvestJob is submitted.

startTime :: Maybe Text

The start of the time-window which will be harvested.

endTime :: Maybe Text

The end of the time-window which will be harvested.

originEndpointId :: Maybe Text

The ID of the OriginEndpoint that the HarvestJob will harvest from. This cannot be changed after the HarvestJob is submitted.

newHarvestJob :: HarvestJob

Create a value of HarvestJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3Destination:HarvestJob', harvestJob_s3Destination - Undocumented member.

$sel:status:HarvestJob', harvestJob_status - The current status of the HarvestJob. Consider setting up a CloudWatch Event to listen for HarvestJobs as they succeed or fail. In the event of failure, the CloudWatch Event will include an explanation of why the HarvestJob failed.

$sel:channelId:HarvestJob', harvestJob_channelId - The ID of the Channel that the HarvestJob will harvest from.

$sel:createdAt:HarvestJob', harvestJob_createdAt - The time the HarvestJob was submitted

$sel:arn:HarvestJob', harvestJob_arn - The Amazon Resource Name (ARN) assigned to the HarvestJob.

$sel:id:HarvestJob', harvestJob_id - The ID of the HarvestJob. The ID must be unique within the region and it cannot be changed after the HarvestJob is submitted.

$sel:startTime:HarvestJob', harvestJob_startTime - The start of the time-window which will be harvested.

$sel:endTime:HarvestJob', harvestJob_endTime - The end of the time-window which will be harvested.

$sel:originEndpointId:HarvestJob', harvestJob_originEndpointId - The ID of the OriginEndpoint that the HarvestJob will harvest from. This cannot be changed after the HarvestJob is submitted.

harvestJob_s3Destination :: Lens' HarvestJob (Maybe S3Destination)

Undocumented member.

harvestJob_status :: Lens' HarvestJob (Maybe Status)

The current status of the HarvestJob. Consider setting up a CloudWatch Event to listen for HarvestJobs as they succeed or fail. In the event of failure, the CloudWatch Event will include an explanation of why the HarvestJob failed.

harvestJob_channelId :: Lens' HarvestJob (Maybe Text)

The ID of the Channel that the HarvestJob will harvest from.

harvestJob_createdAt :: Lens' HarvestJob (Maybe Text)

The time the HarvestJob was submitted

harvestJob_arn :: Lens' HarvestJob (Maybe Text)

The Amazon Resource Name (ARN) assigned to the HarvestJob.

harvestJob_id :: Lens' HarvestJob (Maybe Text)

The ID of the HarvestJob. The ID must be unique within the region and it cannot be changed after the HarvestJob is submitted.

harvestJob_startTime :: Lens' HarvestJob (Maybe Text)

The start of the time-window which will be harvested.

harvestJob_endTime :: Lens' HarvestJob (Maybe Text)

The end of the time-window which will be harvested.

harvestJob_originEndpointId :: Lens' HarvestJob (Maybe Text)

The ID of the OriginEndpoint that the HarvestJob will harvest from. This cannot be changed after the HarvestJob is submitted.