Network.AWS.MediaPackage.Types.HlsManifest

Description

Documentation

data HlsManifest

A HTTP Live Streaming (HLS) manifest configuration.

See: newHlsManifest smart constructor.

Constructors

HlsManifest'

Fields

manifestName :: Maybe Text

An optional short string appended to the end of the OriginEndpoint URL. If not specified, defaults to the manifestName for the OriginEndpoint.

url :: Maybe Text

The URL of the packaged OriginEndpoint for consumption.

playlistType :: Maybe PlaylistType

The HTTP Live Streaming (HLS) playlist type. When either "EVENT" or "VOD" is specified, a corresponding EXT-X-PLAYLIST-TYPE entry will be included in the media playlist.

programDateTimeIntervalSeconds :: Maybe Int

The interval (in seconds) between each EXT-X-PROGRAM-DATE-TIME tag inserted into manifests. Additionally, when an interval is specified ID3Timed Metadata messages will be generated every 5 seconds using the ingest time of the content. If the interval is not specified, or set to 0, then no EXT-X-PROGRAM-DATE-TIME tags will be inserted into manifests and no ID3Timed Metadata messages will be generated. Note that irrespective of this parameter, if any ID3 Timed Metadata is found in HTTP Live Streaming (HLS) input, it will be passed through to HLS output.

adMarkers :: Maybe AdMarkers

This setting controls how ad markers are included in the packaged OriginEndpoint. "NONE" will omit all SCTE-35 ad markers from the output. "PASSTHROUGH" causes the manifest to contain a copy of the SCTE-35 ad markers (comments) taken directly from the input HTTP Live Streaming (HLS) manifest. "SCTE35_ENHANCED" generates ad markers and blackout tags based on SCTE-35 messages in the input source. "DATERANGE" inserts EXT-X-DATERANGE tags to signal ad and program transition events in HLS and CMAF manifests. For this option, you must set a programDateTimeIntervalSeconds value that is greater than 0.

includeIframeOnlyStream :: Maybe Bool

When enabled, an I-Frame only stream will be included in the output.

playlistWindowSeconds :: Maybe Int

Time window (in seconds) contained in each parent manifest.

id :: Text

The ID of the manifest. The ID must be unique within the OriginEndpoint and it cannot be changed after it is created.

newHlsManifest

Arguments

:: Text
$sel:id:HlsManifest'

-> HlsManifest

Create a value of HlsManifest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:manifestName:HlsManifest', hlsManifest_manifestName - An optional short string appended to the end of the OriginEndpoint URL. If not specified, defaults to the manifestName for the OriginEndpoint.

$sel:url:HlsManifest', hlsManifest_url - The URL of the packaged OriginEndpoint for consumption.

$sel:playlistType:HlsManifest', hlsManifest_playlistType - The HTTP Live Streaming (HLS) playlist type. When either "EVENT" or "VOD" is specified, a corresponding EXT-X-PLAYLIST-TYPE entry will be included in the media playlist.

$sel:programDateTimeIntervalSeconds:HlsManifest', hlsManifest_programDateTimeIntervalSeconds - The interval (in seconds) between each EXT-X-PROGRAM-DATE-TIME tag inserted into manifests. Additionally, when an interval is specified ID3Timed Metadata messages will be generated every 5 seconds using the ingest time of the content. If the interval is not specified, or set to 0, then no EXT-X-PROGRAM-DATE-TIME tags will be inserted into manifests and no ID3Timed Metadata messages will be generated. Note that irrespective of this parameter, if any ID3 Timed Metadata is found in HTTP Live Streaming (HLS) input, it will be passed through to HLS output.

$sel:adMarkers:HlsManifest', hlsManifest_adMarkers - This setting controls how ad markers are included in the packaged OriginEndpoint. "NONE" will omit all SCTE-35 ad markers from the output. "PASSTHROUGH" causes the manifest to contain a copy of the SCTE-35 ad markers (comments) taken directly from the input HTTP Live Streaming (HLS) manifest. "SCTE35_ENHANCED" generates ad markers and blackout tags based on SCTE-35 messages in the input source. "DATERANGE" inserts EXT-X-DATERANGE tags to signal ad and program transition events in HLS and CMAF manifests. For this option, you must set a programDateTimeIntervalSeconds value that is greater than 0.

$sel:includeIframeOnlyStream:HlsManifest', hlsManifest_includeIframeOnlyStream - When enabled, an I-Frame only stream will be included in the output.

$sel:playlistWindowSeconds:HlsManifest', hlsManifest_playlistWindowSeconds - Time window (in seconds) contained in each parent manifest.

$sel:id:HlsManifest', hlsManifest_id - The ID of the manifest. The ID must be unique within the OriginEndpoint and it cannot be changed after it is created.

hlsManifest_manifestName :: Lens' HlsManifest (Maybe Text)

An optional short string appended to the end of the OriginEndpoint URL. If not specified, defaults to the manifestName for the OriginEndpoint.

hlsManifest_url :: Lens' HlsManifest (Maybe Text)

The URL of the packaged OriginEndpoint for consumption.

hlsManifest_playlistType :: Lens' HlsManifest (Maybe PlaylistType)

The HTTP Live Streaming (HLS) playlist type. When either "EVENT" or "VOD" is specified, a corresponding EXT-X-PLAYLIST-TYPE entry will be included in the media playlist.

hlsManifest_programDateTimeIntervalSeconds :: Lens' HlsManifest (Maybe Int)

The interval (in seconds) between each EXT-X-PROGRAM-DATE-TIME tag inserted into manifests. Additionally, when an interval is specified ID3Timed Metadata messages will be generated every 5 seconds using the ingest time of the content. If the interval is not specified, or set to 0, then no EXT-X-PROGRAM-DATE-TIME tags will be inserted into manifests and no ID3Timed Metadata messages will be generated. Note that irrespective of this parameter, if any ID3 Timed Metadata is found in HTTP Live Streaming (HLS) input, it will be passed through to HLS output.

hlsManifest_adMarkers :: Lens' HlsManifest (Maybe AdMarkers)

This setting controls how ad markers are included in the packaged OriginEndpoint. "NONE" will omit all SCTE-35 ad markers from the output. "PASSTHROUGH" causes the manifest to contain a copy of the SCTE-35 ad markers (comments) taken directly from the input HTTP Live Streaming (HLS) manifest. "SCTE35_ENHANCED" generates ad markers and blackout tags based on SCTE-35 messages in the input source. "DATERANGE" inserts EXT-X-DATERANGE tags to signal ad and program transition events in HLS and CMAF manifests. For this option, you must set a programDateTimeIntervalSeconds value that is greater than 0.

hlsManifest_includeIframeOnlyStream :: Lens' HlsManifest (Maybe Bool)

When enabled, an I-Frame only stream will be included in the output.

hlsManifest_playlistWindowSeconds :: Lens' HlsManifest (Maybe Int)

Time window (in seconds) contained in each parent manifest.

hlsManifest_id :: Lens' HlsManifest Text

The ID of the manifest. The ID must be unique within the OriginEndpoint and it cannot be changed after it is created.