Network.AWS.MediaPackage.Types.HlsManifestCreateOrUpdateParameters

Description

Documentation

data HlsManifestCreateOrUpdateParameters

A HTTP Live Streaming (HLS) manifest configuration.

See: newHlsManifestCreateOrUpdateParameters smart constructor.

Constructors

HlsManifestCreateOrUpdateParameters'

Fields

adMarkers :: Maybe AdMarkers

This setting controls how ad markers are included in the packaged OriginEndpoint. "NONE" will omit all SCTE-35 ad markers from the output. "PASSTHROUGH" causes the manifest to contain a copy of the SCTE-35 ad markers (comments) taken directly from the input HTTP Live Streaming (HLS) manifest. "SCTE35_ENHANCED" generates ad markers and blackout tags based on SCTE-35 messages in the input source. "DATERANGE" inserts EXT-X-DATERANGE tags to signal ad and program transition events in HLS and CMAF manifests. For this option, you must set a programDateTimeIntervalSeconds value that is greater than 0.

programDateTimeIntervalSeconds :: Maybe Int

The interval (in seconds) between each EXT-X-PROGRAM-DATE-TIME tag inserted into manifests. Additionally, when an interval is specified ID3Timed Metadata messages will be generated every 5 seconds using the ingest time of the content. If the interval is not specified, or set to 0, then no EXT-X-PROGRAM-DATE-TIME tags will be inserted into manifests and no ID3Timed Metadata messages will be generated. Note that irrespective of this parameter, if any ID3 Timed Metadata is found in HTTP Live Streaming (HLS) input, it will be passed through to HLS output.

playlistWindowSeconds :: Maybe Int

Time window (in seconds) contained in each parent manifest.

adTriggers :: Maybe [AdTriggersElement]

includeIframeOnlyStream :: Maybe Bool

When enabled, an I-Frame only stream will be included in the output.

manifestName :: Maybe Text

An optional short string appended to the end of the OriginEndpoint URL. If not specified, defaults to the manifestName for the OriginEndpoint.

adsOnDeliveryRestrictions :: Maybe AdsOnDeliveryRestrictions

playlistType :: Maybe PlaylistType

The HTTP Live Streaming (HLS) playlist type. When either "EVENT" or "VOD" is specified, a corresponding EXT-X-PLAYLIST-TYPE entry will be included in the media playlist.

id :: Text

The ID of the manifest. The ID must be unique within the OriginEndpoint and it cannot be changed after it is created.

newHlsManifestCreateOrUpdateParameters

Arguments

:: Text
$sel:id:HlsManifestCreateOrUpdateParameters'

-> HlsManifestCreateOrUpdateParameters

Create a value of HlsManifestCreateOrUpdateParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:adMarkers:HlsManifestCreateOrUpdateParameters', hlsManifestCreateOrUpdateParameters_adMarkers - This setting controls how ad markers are included in the packaged OriginEndpoint. "NONE" will omit all SCTE-35 ad markers from the output. "PASSTHROUGH" causes the manifest to contain a copy of the SCTE-35 ad markers (comments) taken directly from the input HTTP Live Streaming (HLS) manifest. "SCTE35_ENHANCED" generates ad markers and blackout tags based on SCTE-35 messages in the input source. "DATERANGE" inserts EXT-X-DATERANGE tags to signal ad and program transition events in HLS and CMAF manifests. For this option, you must set a programDateTimeIntervalSeconds value that is greater than 0.

$sel:programDateTimeIntervalSeconds:HlsManifestCreateOrUpdateParameters', hlsManifestCreateOrUpdateParameters_programDateTimeIntervalSeconds - The interval (in seconds) between each EXT-X-PROGRAM-DATE-TIME tag inserted into manifests. Additionally, when an interval is specified ID3Timed Metadata messages will be generated every 5 seconds using the ingest time of the content. If the interval is not specified, or set to 0, then no EXT-X-PROGRAM-DATE-TIME tags will be inserted into manifests and no ID3Timed Metadata messages will be generated. Note that irrespective of this parameter, if any ID3 Timed Metadata is found in HTTP Live Streaming (HLS) input, it will be passed through to HLS output.

$sel:playlistWindowSeconds:HlsManifestCreateOrUpdateParameters', hlsManifestCreateOrUpdateParameters_playlistWindowSeconds - Time window (in seconds) contained in each parent manifest.

$sel:adTriggers:HlsManifestCreateOrUpdateParameters', hlsManifestCreateOrUpdateParameters_adTriggers - Undocumented member.

$sel:includeIframeOnlyStream:HlsManifestCreateOrUpdateParameters', hlsManifestCreateOrUpdateParameters_includeIframeOnlyStream - When enabled, an I-Frame only stream will be included in the output.

$sel:manifestName:HlsManifestCreateOrUpdateParameters', hlsManifestCreateOrUpdateParameters_manifestName - An optional short string appended to the end of the OriginEndpoint URL. If not specified, defaults to the manifestName for the OriginEndpoint.

$sel:adsOnDeliveryRestrictions:HlsManifestCreateOrUpdateParameters', hlsManifestCreateOrUpdateParameters_adsOnDeliveryRestrictions - Undocumented member.

$sel:playlistType:HlsManifestCreateOrUpdateParameters', hlsManifestCreateOrUpdateParameters_playlistType - The HTTP Live Streaming (HLS) playlist type. When either "EVENT" or "VOD" is specified, a corresponding EXT-X-PLAYLIST-TYPE entry will be included in the media playlist.

$sel:id:HlsManifestCreateOrUpdateParameters', hlsManifestCreateOrUpdateParameters_id - The ID of the manifest. The ID must be unique within the OriginEndpoint and it cannot be changed after it is created.

hlsManifestCreateOrUpdateParameters_adMarkers :: Lens' HlsManifestCreateOrUpdateParameters (Maybe AdMarkers)

This setting controls how ad markers are included in the packaged OriginEndpoint. "NONE" will omit all SCTE-35 ad markers from the output. "PASSTHROUGH" causes the manifest to contain a copy of the SCTE-35 ad markers (comments) taken directly from the input HTTP Live Streaming (HLS) manifest. "SCTE35_ENHANCED" generates ad markers and blackout tags based on SCTE-35 messages in the input source. "DATERANGE" inserts EXT-X-DATERANGE tags to signal ad and program transition events in HLS and CMAF manifests. For this option, you must set a programDateTimeIntervalSeconds value that is greater than 0.

hlsManifestCreateOrUpdateParameters_programDateTimeIntervalSeconds :: Lens' HlsManifestCreateOrUpdateParameters (Maybe Int)

The interval (in seconds) between each EXT-X-PROGRAM-DATE-TIME tag inserted into manifests. Additionally, when an interval is specified ID3Timed Metadata messages will be generated every 5 seconds using the ingest time of the content. If the interval is not specified, or set to 0, then no EXT-X-PROGRAM-DATE-TIME tags will be inserted into manifests and no ID3Timed Metadata messages will be generated. Note that irrespective of this parameter, if any ID3 Timed Metadata is found in HTTP Live Streaming (HLS) input, it will be passed through to HLS output.

hlsManifestCreateOrUpdateParameters_playlistWindowSeconds :: Lens' HlsManifestCreateOrUpdateParameters (Maybe Int)

Time window (in seconds) contained in each parent manifest.

hlsManifestCreateOrUpdateParameters_adTriggers :: Lens' HlsManifestCreateOrUpdateParameters (Maybe [AdTriggersElement])

Undocumented member.

hlsManifestCreateOrUpdateParameters_includeIframeOnlyStream :: Lens' HlsManifestCreateOrUpdateParameters (Maybe Bool)

When enabled, an I-Frame only stream will be included in the output.

hlsManifestCreateOrUpdateParameters_manifestName :: Lens' HlsManifestCreateOrUpdateParameters (Maybe Text)

An optional short string appended to the end of the OriginEndpoint URL. If not specified, defaults to the manifestName for the OriginEndpoint.

hlsManifestCreateOrUpdateParameters_adsOnDeliveryRestrictions :: Lens' HlsManifestCreateOrUpdateParameters (Maybe AdsOnDeliveryRestrictions)

Undocumented member.

hlsManifestCreateOrUpdateParameters_playlistType :: Lens' HlsManifestCreateOrUpdateParameters (Maybe PlaylistType)

The HTTP Live Streaming (HLS) playlist type. When either "EVENT" or "VOD" is specified, a corresponding EXT-X-PLAYLIST-TYPE entry will be included in the media playlist.

hlsManifestCreateOrUpdateParameters_id :: Lens' HlsManifestCreateOrUpdateParameters Text

The ID of the manifest. The ID must be unique within the OriginEndpoint and it cannot be changed after it is created.