Network.AWS.MediaPackage.Types.HlsPackage

Description

Documentation

data HlsPackage

An HTTP Live Streaming (HLS) packaging configuration.

See: newHlsPackage smart constructor.

Constructors

HlsPackage'

Fields

adsOnDeliveryRestrictions :: Maybe AdsOnDeliveryRestrictions

useAudioRenditionGroup :: Maybe Bool

When enabled, audio streams will be placed in rendition groups in the output.

includeDvbSubtitles :: Maybe Bool

When enabled, MediaPackage passes through digital video broadcasting (DVB) subtitles into the output.

playlistType :: Maybe PlaylistType

The HTTP Live Streaming (HLS) playlist type. When either "EVENT" or "VOD" is specified, a corresponding EXT-X-PLAYLIST-TYPE entry will be included in the media playlist.

segmentDurationSeconds :: Maybe Int

Duration (in seconds) of each fragment. Actual fragments will be rounded to the nearest multiple of the source fragment duration.

programDateTimeIntervalSeconds :: Maybe Int

The interval (in seconds) between each EXT-X-PROGRAM-DATE-TIME tag inserted into manifests. Additionally, when an interval is specified ID3Timed Metadata messages will be generated every 5 seconds using the ingest time of the content. If the interval is not specified, or set to 0, then no EXT-X-PROGRAM-DATE-TIME tags will be inserted into manifests and no ID3Timed Metadata messages will be generated. Note that irrespective of this parameter, if any ID3 Timed Metadata is found in HTTP Live Streaming (HLS) input, it will be passed through to HLS output.

streamSelection :: Maybe StreamSelection

adMarkers :: Maybe AdMarkers

This setting controls how ad markers are included in the packaged OriginEndpoint. "NONE" will omit all SCTE-35 ad markers from the output. "PASSTHROUGH" causes the manifest to contain a copy of the SCTE-35 ad markers (comments) taken directly from the input HTTP Live Streaming (HLS) manifest. "SCTE35_ENHANCED" generates ad markers and blackout tags based on SCTE-35 messages in the input source. "DATERANGE" inserts EXT-X-DATERANGE tags to signal ad and program transition events in HLS and CMAF manifests. For this option, you must set a programDateTimeIntervalSeconds value that is greater than 0.

encryption :: Maybe HlsEncryption

includeIframeOnlyStream :: Maybe Bool

When enabled, an I-Frame only stream will be included in the output.

adTriggers :: Maybe [AdTriggersElement]

playlistWindowSeconds :: Maybe Int

Time window (in seconds) contained in each parent manifest.

newHlsPackage :: HlsPackage

Create a value of HlsPackage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:adsOnDeliveryRestrictions:HlsPackage', hlsPackage_adsOnDeliveryRestrictions - Undocumented member.

$sel:useAudioRenditionGroup:HlsPackage', hlsPackage_useAudioRenditionGroup - When enabled, audio streams will be placed in rendition groups in the output.

$sel:includeDvbSubtitles:HlsPackage', hlsPackage_includeDvbSubtitles - When enabled, MediaPackage passes through digital video broadcasting (DVB) subtitles into the output.

$sel:playlistType:HlsPackage', hlsPackage_playlistType - The HTTP Live Streaming (HLS) playlist type. When either "EVENT" or "VOD" is specified, a corresponding EXT-X-PLAYLIST-TYPE entry will be included in the media playlist.

$sel:segmentDurationSeconds:HlsPackage', hlsPackage_segmentDurationSeconds - Duration (in seconds) of each fragment. Actual fragments will be rounded to the nearest multiple of the source fragment duration.

$sel:programDateTimeIntervalSeconds:HlsPackage', hlsPackage_programDateTimeIntervalSeconds - The interval (in seconds) between each EXT-X-PROGRAM-DATE-TIME tag inserted into manifests. Additionally, when an interval is specified ID3Timed Metadata messages will be generated every 5 seconds using the ingest time of the content. If the interval is not specified, or set to 0, then no EXT-X-PROGRAM-DATE-TIME tags will be inserted into manifests and no ID3Timed Metadata messages will be generated. Note that irrespective of this parameter, if any ID3 Timed Metadata is found in HTTP Live Streaming (HLS) input, it will be passed through to HLS output.

$sel:streamSelection:HlsPackage', hlsPackage_streamSelection - Undocumented member.

$sel:adMarkers:HlsPackage', hlsPackage_adMarkers - This setting controls how ad markers are included in the packaged OriginEndpoint. "NONE" will omit all SCTE-35 ad markers from the output. "PASSTHROUGH" causes the manifest to contain a copy of the SCTE-35 ad markers (comments) taken directly from the input HTTP Live Streaming (HLS) manifest. "SCTE35_ENHANCED" generates ad markers and blackout tags based on SCTE-35 messages in the input source. "DATERANGE" inserts EXT-X-DATERANGE tags to signal ad and program transition events in HLS and CMAF manifests. For this option, you must set a programDateTimeIntervalSeconds value that is greater than 0.

$sel:encryption:HlsPackage', hlsPackage_encryption - Undocumented member.

$sel:includeIframeOnlyStream:HlsPackage', hlsPackage_includeIframeOnlyStream - When enabled, an I-Frame only stream will be included in the output.

$sel:adTriggers:HlsPackage', hlsPackage_adTriggers - Undocumented member.

$sel:playlistWindowSeconds:HlsPackage', hlsPackage_playlistWindowSeconds - Time window (in seconds) contained in each parent manifest.

hlsPackage_adsOnDeliveryRestrictions :: Lens' HlsPackage (Maybe AdsOnDeliveryRestrictions)

Undocumented member.

hlsPackage_useAudioRenditionGroup :: Lens' HlsPackage (Maybe Bool)

When enabled, audio streams will be placed in rendition groups in the output.

hlsPackage_includeDvbSubtitles :: Lens' HlsPackage (Maybe Bool)

When enabled, MediaPackage passes through digital video broadcasting (DVB) subtitles into the output.

hlsPackage_playlistType :: Lens' HlsPackage (Maybe PlaylistType)

The HTTP Live Streaming (HLS) playlist type. When either "EVENT" or "VOD" is specified, a corresponding EXT-X-PLAYLIST-TYPE entry will be included in the media playlist.

hlsPackage_segmentDurationSeconds :: Lens' HlsPackage (Maybe Int)

Duration (in seconds) of each fragment. Actual fragments will be rounded to the nearest multiple of the source fragment duration.

hlsPackage_programDateTimeIntervalSeconds :: Lens' HlsPackage (Maybe Int)

The interval (in seconds) between each EXT-X-PROGRAM-DATE-TIME tag inserted into manifests. Additionally, when an interval is specified ID3Timed Metadata messages will be generated every 5 seconds using the ingest time of the content. If the interval is not specified, or set to 0, then no EXT-X-PROGRAM-DATE-TIME tags will be inserted into manifests and no ID3Timed Metadata messages will be generated. Note that irrespective of this parameter, if any ID3 Timed Metadata is found in HTTP Live Streaming (HLS) input, it will be passed through to HLS output.

hlsPackage_streamSelection :: Lens' HlsPackage (Maybe StreamSelection)

Undocumented member.

hlsPackage_adMarkers :: Lens' HlsPackage (Maybe AdMarkers)

This setting controls how ad markers are included in the packaged OriginEndpoint. "NONE" will omit all SCTE-35 ad markers from the output. "PASSTHROUGH" causes the manifest to contain a copy of the SCTE-35 ad markers (comments) taken directly from the input HTTP Live Streaming (HLS) manifest. "SCTE35_ENHANCED" generates ad markers and blackout tags based on SCTE-35 messages in the input source. "DATERANGE" inserts EXT-X-DATERANGE tags to signal ad and program transition events in HLS and CMAF manifests. For this option, you must set a programDateTimeIntervalSeconds value that is greater than 0.

hlsPackage_encryption :: Lens' HlsPackage (Maybe HlsEncryption)

Undocumented member.

hlsPackage_includeIframeOnlyStream :: Lens' HlsPackage (Maybe Bool)

When enabled, an I-Frame only stream will be included in the output.

hlsPackage_adTriggers :: Lens' HlsPackage (Maybe [AdTriggersElement])

Undocumented member.

hlsPackage_playlistWindowSeconds :: Lens' HlsPackage (Maybe Int)

Time window (in seconds) contained in each parent manifest.