Network.AWS.MediaPackage.Types.S3Destination

Description

Documentation

data S3Destination

Configuration parameters for where in an S3 bucket to place the harvested content

See: newS3Destination smart constructor.

Constructors

S3Destination'

Fields

manifestKey :: Text

The key in the specified S3 bucket where the harvested top-level manifest will be placed.

bucketName :: Text

The name of an S3 bucket within which harvested content will be exported

roleArn :: Text

The IAM role used to write to the specified S3 bucket

newS3Destination

Arguments

:: Text
$sel:manifestKey:S3Destination'

-> Text
$sel:bucketName:S3Destination'

-> Text
$sel:roleArn:S3Destination'

-> S3Destination

Create a value of S3Destination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:manifestKey:S3Destination', s3Destination_manifestKey - The key in the specified S3 bucket where the harvested top-level manifest will be placed.

$sel:bucketName:S3Destination', s3Destination_bucketName - The name of an S3 bucket within which harvested content will be exported

$sel:roleArn:S3Destination', s3Destination_roleArn - The IAM role used to write to the specified S3 bucket

s3Destination_manifestKey :: Lens' S3Destination Text

The key in the specified S3 bucket where the harvested top-level manifest will be placed.

s3Destination_bucketName :: Lens' S3Destination Text

The name of an S3 bucket within which harvested content will be exported

s3Destination_roleArn :: Lens' S3Destination Text

The IAM role used to write to the specified S3 bucket