Network.AWS.MediaPackage.Types.StreamSelection

Description

Documentation

data StreamSelection

A StreamSelection configuration.

See: newStreamSelection smart constructor.

Constructors

StreamSelection'

Fields

streamOrder :: Maybe StreamOrder

A directive that determines the order of streams in the output.

minVideoBitsPerSecond :: Maybe Int

The minimum video bitrate (bps) to include in output.

maxVideoBitsPerSecond :: Maybe Int

The maximum video bitrate (bps) to include in output.

newStreamSelection :: StreamSelection

Create a value of StreamSelection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamOrder:StreamSelection', streamSelection_streamOrder - A directive that determines the order of streams in the output.

$sel:minVideoBitsPerSecond:StreamSelection', streamSelection_minVideoBitsPerSecond - The minimum video bitrate (bps) to include in output.

$sel:maxVideoBitsPerSecond:StreamSelection', streamSelection_maxVideoBitsPerSecond - The maximum video bitrate (bps) to include in output.

streamSelection_streamOrder :: Lens' StreamSelection (Maybe StreamOrder)

A directive that determines the order of streams in the output.

streamSelection_minVideoBitsPerSecond :: Lens' StreamSelection (Maybe Int)

The minimum video bitrate (bps) to include in output.

streamSelection_maxVideoBitsPerSecond :: Lens' StreamSelection (Maybe Int)

The maximum video bitrate (bps) to include in output.