Network.AWS.MediaPackage.UpdateOriginEndpoint

Description

Updates an existing OriginEndpoint.

Creating a Request

data UpdateOriginEndpoint

Configuration parameters used to update an existing OriginEndpoint.

See: newUpdateOriginEndpoint smart constructor.

Constructors

UpdateOriginEndpoint'

Fields

dashPackage :: Maybe DashPackage

origination :: Maybe Origination

Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

startoverWindowSeconds :: Maybe Int

Maximum duration (in seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

cmafPackage :: Maybe CmafPackageCreateOrUpdateParameters

manifestName :: Maybe Text

A short string that will be appended to the end of the Endpoint URL.

whitelist :: Maybe [Text]

A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

mssPackage :: Maybe MssPackage

description :: Maybe Text

A short text description of the OriginEndpoint.

timeDelaySeconds :: Maybe Int

Amount of delay (in seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

authorization :: Maybe Authorization

hlsPackage :: Maybe HlsPackage

id :: Text

The ID of the OriginEndpoint to update.

newUpdateOriginEndpoint

Arguments

:: Text
$sel:id:UpdateOriginEndpoint'

-> UpdateOriginEndpoint

Create a value of UpdateOriginEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dashPackage:UpdateOriginEndpoint', updateOriginEndpoint_dashPackage - Undocumented member.

$sel:origination:UpdateOriginEndpoint', updateOriginEndpoint_origination - Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

$sel:startoverWindowSeconds:UpdateOriginEndpoint', updateOriginEndpoint_startoverWindowSeconds - Maximum duration (in seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

$sel:cmafPackage:UpdateOriginEndpoint', updateOriginEndpoint_cmafPackage - Undocumented member.

$sel:manifestName:UpdateOriginEndpoint', updateOriginEndpoint_manifestName - A short string that will be appended to the end of the Endpoint URL.

$sel:whitelist:UpdateOriginEndpoint', updateOriginEndpoint_whitelist - A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

$sel:mssPackage:UpdateOriginEndpoint', updateOriginEndpoint_mssPackage - Undocumented member.

$sel:description:UpdateOriginEndpoint', updateOriginEndpoint_description - A short text description of the OriginEndpoint.

$sel:timeDelaySeconds:UpdateOriginEndpoint', updateOriginEndpoint_timeDelaySeconds - Amount of delay (in seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

$sel:authorization:UpdateOriginEndpoint', updateOriginEndpoint_authorization - Undocumented member.

$sel:hlsPackage:UpdateOriginEndpoint', updateOriginEndpoint_hlsPackage - Undocumented member.

$sel:id:UpdateOriginEndpoint', updateOriginEndpoint_id - The ID of the OriginEndpoint to update.

Request Lenses

updateOriginEndpoint_dashPackage :: Lens' UpdateOriginEndpoint (Maybe DashPackage)

Undocumented member.

updateOriginEndpoint_origination :: Lens' UpdateOriginEndpoint (Maybe Origination)

Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

updateOriginEndpoint_startoverWindowSeconds :: Lens' UpdateOriginEndpoint (Maybe Int)

Maximum duration (in seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

updateOriginEndpoint_cmafPackage :: Lens' UpdateOriginEndpoint (Maybe CmafPackageCreateOrUpdateParameters)

Undocumented member.

updateOriginEndpoint_manifestName :: Lens' UpdateOriginEndpoint (Maybe Text)

A short string that will be appended to the end of the Endpoint URL.

updateOriginEndpoint_whitelist :: Lens' UpdateOriginEndpoint (Maybe [Text])

A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

updateOriginEndpoint_mssPackage :: Lens' UpdateOriginEndpoint (Maybe MssPackage)

Undocumented member.

updateOriginEndpoint_description :: Lens' UpdateOriginEndpoint (Maybe Text)

A short text description of the OriginEndpoint.

updateOriginEndpoint_timeDelaySeconds :: Lens' UpdateOriginEndpoint (Maybe Int)

Amount of delay (in seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

updateOriginEndpoint_authorization :: Lens' UpdateOriginEndpoint (Maybe Authorization)

Undocumented member.

updateOriginEndpoint_hlsPackage :: Lens' UpdateOriginEndpoint (Maybe HlsPackage)

Undocumented member.

updateOriginEndpoint_id :: Lens' UpdateOriginEndpoint Text

The ID of the OriginEndpoint to update.

Destructuring the Response

data UpdateOriginEndpointResponse

See: newUpdateOriginEndpointResponse smart constructor.

Constructors

UpdateOriginEndpointResponse'

Fields

dashPackage :: Maybe DashPackage

origination :: Maybe Origination

Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

startoverWindowSeconds :: Maybe Int

Maximum duration (seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

channelId :: Maybe Text

The ID of the Channel the OriginEndpoint is associated with.

cmafPackage :: Maybe CmafPackage

manifestName :: Maybe Text

A short string appended to the end of the OriginEndpoint URL.

arn :: Maybe Text

The Amazon Resource Name (ARN) assigned to the OriginEndpoint.

id :: Maybe Text

The ID of the OriginEndpoint.

whitelist :: Maybe [Text]

A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

mssPackage :: Maybe MssPackage

tags :: Maybe (HashMap Text Text)

description :: Maybe Text

A short text description of the OriginEndpoint.

url :: Maybe Text

The URL of the packaged OriginEndpoint for consumption.

timeDelaySeconds :: Maybe Int

Amount of delay (seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

authorization :: Maybe Authorization

hlsPackage :: Maybe HlsPackage

httpStatus :: Int

The response's http status code.

newUpdateOriginEndpointResponse

Arguments

:: Int
$sel:httpStatus:UpdateOriginEndpointResponse'

-> UpdateOriginEndpointResponse

Create a value of UpdateOriginEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dashPackage:UpdateOriginEndpoint', updateOriginEndpointResponse_dashPackage - Undocumented member.

$sel:origination:UpdateOriginEndpoint', updateOriginEndpointResponse_origination - Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

$sel:startoverWindowSeconds:UpdateOriginEndpoint', updateOriginEndpointResponse_startoverWindowSeconds - Maximum duration (seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

$sel:channelId:UpdateOriginEndpointResponse', updateOriginEndpointResponse_channelId - The ID of the Channel the OriginEndpoint is associated with.

$sel:cmafPackage:UpdateOriginEndpoint', updateOriginEndpointResponse_cmafPackage - Undocumented member.

$sel:manifestName:UpdateOriginEndpoint', updateOriginEndpointResponse_manifestName - A short string appended to the end of the OriginEndpoint URL.

$sel:arn:UpdateOriginEndpointResponse', updateOriginEndpointResponse_arn - The Amazon Resource Name (ARN) assigned to the OriginEndpoint.

$sel:id:UpdateOriginEndpoint', updateOriginEndpointResponse_id - The ID of the OriginEndpoint.

$sel:whitelist:UpdateOriginEndpoint', updateOriginEndpointResponse_whitelist - A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

$sel:mssPackage:UpdateOriginEndpoint', updateOriginEndpointResponse_mssPackage - Undocumented member.

$sel:tags:UpdateOriginEndpointResponse', updateOriginEndpointResponse_tags - Undocumented member.

$sel:description:UpdateOriginEndpoint', updateOriginEndpointResponse_description - A short text description of the OriginEndpoint.

$sel:url:UpdateOriginEndpointResponse', updateOriginEndpointResponse_url - The URL of the packaged OriginEndpoint for consumption.

$sel:timeDelaySeconds:UpdateOriginEndpoint', updateOriginEndpointResponse_timeDelaySeconds - Amount of delay (seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

$sel:authorization:UpdateOriginEndpoint', updateOriginEndpointResponse_authorization - Undocumented member.

$sel:hlsPackage:UpdateOriginEndpoint', updateOriginEndpointResponse_hlsPackage - Undocumented member.

$sel:httpStatus:UpdateOriginEndpointResponse', updateOriginEndpointResponse_httpStatus - The response's http status code.

Response Lenses

updateOriginEndpointResponse_dashPackage :: Lens' UpdateOriginEndpointResponse (Maybe DashPackage)

Undocumented member.

updateOriginEndpointResponse_origination :: Lens' UpdateOriginEndpointResponse (Maybe Origination)

Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

updateOriginEndpointResponse_startoverWindowSeconds :: Lens' UpdateOriginEndpointResponse (Maybe Int)

Maximum duration (seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

updateOriginEndpointResponse_channelId :: Lens' UpdateOriginEndpointResponse (Maybe Text)

The ID of the Channel the OriginEndpoint is associated with.

updateOriginEndpointResponse_cmafPackage :: Lens' UpdateOriginEndpointResponse (Maybe CmafPackage)

Undocumented member.

updateOriginEndpointResponse_manifestName :: Lens' UpdateOriginEndpointResponse (Maybe Text)

A short string appended to the end of the OriginEndpoint URL.

updateOriginEndpointResponse_arn :: Lens' UpdateOriginEndpointResponse (Maybe Text)

The Amazon Resource Name (ARN) assigned to the OriginEndpoint.

updateOriginEndpointResponse_id :: Lens' UpdateOriginEndpointResponse (Maybe Text)

The ID of the OriginEndpoint.

updateOriginEndpointResponse_whitelist :: Lens' UpdateOriginEndpointResponse (Maybe [Text])

A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

updateOriginEndpointResponse_mssPackage :: Lens' UpdateOriginEndpointResponse (Maybe MssPackage)

Undocumented member.

updateOriginEndpointResponse_tags :: Lens' UpdateOriginEndpointResponse (Maybe (HashMap Text Text))

Undocumented member.

updateOriginEndpointResponse_description :: Lens' UpdateOriginEndpointResponse (Maybe Text)

A short text description of the OriginEndpoint.

updateOriginEndpointResponse_url :: Lens' UpdateOriginEndpointResponse (Maybe Text)

The URL of the packaged OriginEndpoint for consumption.

updateOriginEndpointResponse_timeDelaySeconds :: Lens' UpdateOriginEndpointResponse (Maybe Int)

Amount of delay (seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

updateOriginEndpointResponse_authorization :: Lens' UpdateOriginEndpointResponse (Maybe Authorization)

Undocumented member.

updateOriginEndpointResponse_hlsPackage :: Lens' UpdateOriginEndpointResponse (Maybe HlsPackage)

Undocumented member.

updateOriginEndpointResponse_httpStatus :: Lens' UpdateOriginEndpointResponse Int

The response's http status code.