Network.AWS.MediaPackageVOD.DescribePackagingConfiguration

Description

Returns a description of a MediaPackage VOD PackagingConfiguration resource.

Creating a Request

data DescribePackagingConfiguration

See: newDescribePackagingConfiguration smart constructor.

Constructors

DescribePackagingConfiguration'

Fields

id :: Text

The ID of a MediaPackage VOD PackagingConfiguration resource.

newDescribePackagingConfiguration

Arguments

:: Text
$sel:id:DescribePackagingConfiguration'

-> DescribePackagingConfiguration

Create a value of DescribePackagingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DescribePackagingConfiguration', describePackagingConfiguration_id - The ID of a MediaPackage VOD PackagingConfiguration resource.

Request Lenses

describePackagingConfiguration_id :: Lens' DescribePackagingConfiguration Text

The ID of a MediaPackage VOD PackagingConfiguration resource.

Destructuring the Response

data DescribePackagingConfigurationResponse

See: newDescribePackagingConfigurationResponse smart constructor.

Constructors

DescribePackagingConfigurationResponse'

Fields

hlsPackage :: Maybe HlsPackage

arn :: Maybe Text

The ARN of the PackagingConfiguration.

packagingGroupId :: Maybe Text

The ID of a PackagingGroup.

dashPackage :: Maybe DashPackage

mssPackage :: Maybe MssPackage

id :: Maybe Text

The ID of the PackagingConfiguration.

cmafPackage :: Maybe CmafPackage

tags :: Maybe (HashMap Text Text)

httpStatus :: Int

The response's http status code.

newDescribePackagingConfigurationResponse

Arguments

:: Int
$sel:httpStatus:DescribePackagingConfigurationResponse'

-> DescribePackagingConfigurationResponse

Create a value of DescribePackagingConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hlsPackage:DescribePackagingConfigurationResponse', describePackagingConfigurationResponse_hlsPackage - Undocumented member.

$sel:arn:DescribePackagingConfigurationResponse', describePackagingConfigurationResponse_arn - The ARN of the PackagingConfiguration.

$sel:packagingGroupId:DescribePackagingConfigurationResponse', describePackagingConfigurationResponse_packagingGroupId - The ID of a PackagingGroup.

$sel:dashPackage:DescribePackagingConfigurationResponse', describePackagingConfigurationResponse_dashPackage - Undocumented member.

$sel:mssPackage:DescribePackagingConfigurationResponse', describePackagingConfigurationResponse_mssPackage - Undocumented member.

$sel:id:DescribePackagingConfiguration', describePackagingConfigurationResponse_id - The ID of the PackagingConfiguration.

$sel:cmafPackage:DescribePackagingConfigurationResponse', describePackagingConfigurationResponse_cmafPackage - Undocumented member.

$sel:tags:DescribePackagingConfigurationResponse', describePackagingConfigurationResponse_tags - Undocumented member.

$sel:httpStatus:DescribePackagingConfigurationResponse', describePackagingConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

describePackagingConfigurationResponse_hlsPackage :: Lens' DescribePackagingConfigurationResponse (Maybe HlsPackage)

Undocumented member.

describePackagingConfigurationResponse_arn :: Lens' DescribePackagingConfigurationResponse (Maybe Text)

The ARN of the PackagingConfiguration.

describePackagingConfigurationResponse_packagingGroupId :: Lens' DescribePackagingConfigurationResponse (Maybe Text)

The ID of a PackagingGroup.

describePackagingConfigurationResponse_dashPackage :: Lens' DescribePackagingConfigurationResponse (Maybe DashPackage)

Undocumented member.

describePackagingConfigurationResponse_mssPackage :: Lens' DescribePackagingConfigurationResponse (Maybe MssPackage)

Undocumented member.

describePackagingConfigurationResponse_id :: Lens' DescribePackagingConfigurationResponse (Maybe Text)

The ID of the PackagingConfiguration.

describePackagingConfigurationResponse_cmafPackage :: Lens' DescribePackagingConfigurationResponse (Maybe CmafPackage)

Undocumented member.

describePackagingConfigurationResponse_tags :: Lens' DescribePackagingConfigurationResponse (Maybe (HashMap Text Text))

Undocumented member.

describePackagingConfigurationResponse_httpStatus :: Lens' DescribePackagingConfigurationResponse Int

The response's http status code.