Network.AWS.MediaPackageVOD.DescribePackagingGroup

Description

Returns a description of a MediaPackage VOD PackagingGroup resource.

Creating a Request

data DescribePackagingGroup

See: newDescribePackagingGroup smart constructor.

Constructors

DescribePackagingGroup'

Fields

id :: Text

The ID of a MediaPackage VOD PackagingGroup resource.

newDescribePackagingGroup

Arguments

:: Text
$sel:id:DescribePackagingGroup'

-> DescribePackagingGroup

Create a value of DescribePackagingGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:DescribePackagingGroup', describePackagingGroup_id - The ID of a MediaPackage VOD PackagingGroup resource.

Request Lenses

describePackagingGroup_id :: Lens' DescribePackagingGroup Text

The ID of a MediaPackage VOD PackagingGroup resource.

Destructuring the Response

data DescribePackagingGroupResponse

See: newDescribePackagingGroupResponse smart constructor.

Constructors

DescribePackagingGroupResponse'

Fields

arn :: Maybe Text

The ARN of the PackagingGroup.

authorization :: Maybe Authorization

domainName :: Maybe Text

The fully qualified domain name for Assets in the PackagingGroup.

id :: Maybe Text

The ID of the PackagingGroup.

egressAccessLogs :: Maybe EgressAccessLogs

tags :: Maybe (HashMap Text Text)

httpStatus :: Int

The response's http status code.

newDescribePackagingGroupResponse

Arguments

:: Int
$sel:httpStatus:DescribePackagingGroupResponse'

-> DescribePackagingGroupResponse

Create a value of DescribePackagingGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:DescribePackagingGroupResponse', describePackagingGroupResponse_arn - The ARN of the PackagingGroup.

$sel:authorization:DescribePackagingGroupResponse', describePackagingGroupResponse_authorization - Undocumented member.

$sel:domainName:DescribePackagingGroupResponse', describePackagingGroupResponse_domainName - The fully qualified domain name for Assets in the PackagingGroup.

$sel:id:DescribePackagingGroup', describePackagingGroupResponse_id - The ID of the PackagingGroup.

$sel:egressAccessLogs:DescribePackagingGroupResponse', describePackagingGroupResponse_egressAccessLogs - Undocumented member.

$sel:tags:DescribePackagingGroupResponse', describePackagingGroupResponse_tags - Undocumented member.

$sel:httpStatus:DescribePackagingGroupResponse', describePackagingGroupResponse_httpStatus - The response's http status code.

Response Lenses

describePackagingGroupResponse_arn :: Lens' DescribePackagingGroupResponse (Maybe Text)

The ARN of the PackagingGroup.

describePackagingGroupResponse_authorization :: Lens' DescribePackagingGroupResponse (Maybe Authorization)

Undocumented member.

describePackagingGroupResponse_domainName :: Lens' DescribePackagingGroupResponse (Maybe Text)

The fully qualified domain name for Assets in the PackagingGroup.

describePackagingGroupResponse_id :: Lens' DescribePackagingGroupResponse (Maybe Text)

The ID of the PackagingGroup.

describePackagingGroupResponse_egressAccessLogs :: Lens' DescribePackagingGroupResponse (Maybe EgressAccessLogs)

Undocumented member.

describePackagingGroupResponse_tags :: Lens' DescribePackagingGroupResponse (Maybe (HashMap Text Text))

Undocumented member.

describePackagingGroupResponse_httpStatus :: Lens' DescribePackagingGroupResponse Int

The response's http status code.