Network.AWS.MediaPackageVOD.Lens

Description

Operations

CreatePackagingGroup

createPackagingGroup_authorization :: Lens' CreatePackagingGroup (Maybe Authorization)

Undocumented member.

createPackagingGroup_egressAccessLogs :: Lens' CreatePackagingGroup (Maybe EgressAccessLogs)

Undocumented member.

createPackagingGroup_tags :: Lens' CreatePackagingGroup (Maybe (HashMap Text Text))

Undocumented member.

createPackagingGroup_id :: Lens' CreatePackagingGroup Text

The ID of the PackagingGroup.

createPackagingGroupResponse_arn :: Lens' CreatePackagingGroupResponse (Maybe Text)

The ARN of the PackagingGroup.

createPackagingGroupResponse_authorization :: Lens' CreatePackagingGroupResponse (Maybe Authorization)

Undocumented member.

createPackagingGroupResponse_domainName :: Lens' CreatePackagingGroupResponse (Maybe Text)

The fully qualified domain name for Assets in the PackagingGroup.

createPackagingGroupResponse_id :: Lens' CreatePackagingGroupResponse (Maybe Text)

The ID of the PackagingGroup.

createPackagingGroupResponse_egressAccessLogs :: Lens' CreatePackagingGroupResponse (Maybe EgressAccessLogs)

Undocumented member.

createPackagingGroupResponse_tags :: Lens' CreatePackagingGroupResponse (Maybe (HashMap Text Text))

Undocumented member.

createPackagingGroupResponse_httpStatus :: Lens' CreatePackagingGroupResponse Int

The response's http status code.

ConfigureLogs

configureLogs_egressAccessLogs :: Lens' ConfigureLogs (Maybe EgressAccessLogs)

Undocumented member.

configureLogs_id :: Lens' ConfigureLogs Text

The ID of a MediaPackage VOD PackagingGroup resource.

configureLogsResponse_arn :: Lens' ConfigureLogsResponse (Maybe Text)

The ARN of the PackagingGroup.

configureLogsResponse_authorization :: Lens' ConfigureLogsResponse (Maybe Authorization)

Undocumented member.

configureLogsResponse_domainName :: Lens' ConfigureLogsResponse (Maybe Text)

The fully qualified domain name for Assets in the PackagingGroup.

configureLogsResponse_id :: Lens' ConfigureLogsResponse (Maybe Text)

The ID of the PackagingGroup.

configureLogsResponse_egressAccessLogs :: Lens' ConfigureLogsResponse (Maybe EgressAccessLogs)

Undocumented member.

configureLogsResponse_tags :: Lens' ConfigureLogsResponse (Maybe (HashMap Text Text))

Undocumented member.

configureLogsResponse_httpStatus :: Lens' ConfigureLogsResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) for the resource. You can get this from the response to any request to the resource.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

A collection of tags associated with a resource

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

DescribePackagingGroup

describePackagingGroup_id :: Lens' DescribePackagingGroup Text

The ID of a MediaPackage VOD PackagingGroup resource.

describePackagingGroupResponse_arn :: Lens' DescribePackagingGroupResponse (Maybe Text)

The ARN of the PackagingGroup.

describePackagingGroupResponse_authorization :: Lens' DescribePackagingGroupResponse (Maybe Authorization)

Undocumented member.

describePackagingGroupResponse_domainName :: Lens' DescribePackagingGroupResponse (Maybe Text)

The fully qualified domain name for Assets in the PackagingGroup.

describePackagingGroupResponse_id :: Lens' DescribePackagingGroupResponse (Maybe Text)

The ID of the PackagingGroup.

describePackagingGroupResponse_egressAccessLogs :: Lens' DescribePackagingGroupResponse (Maybe EgressAccessLogs)

Undocumented member.

describePackagingGroupResponse_tags :: Lens' DescribePackagingGroupResponse (Maybe (HashMap Text Text))

Undocumented member.

describePackagingGroupResponse_httpStatus :: Lens' DescribePackagingGroupResponse Int

The response's http status code.

DescribeAsset

describeAsset_id :: Lens' DescribeAsset Text

The ID of an MediaPackage VOD Asset resource.

describeAssetResponse_resourceId :: Lens' DescribeAssetResponse (Maybe Text)

The resource ID to include in SPEKE key requests.

describeAssetResponse_arn :: Lens' DescribeAssetResponse (Maybe Text)

The ARN of the Asset.

describeAssetResponse_createdAt :: Lens' DescribeAssetResponse (Maybe Text)

The time the Asset was initially submitted for Ingest.

describeAssetResponse_packagingGroupId :: Lens' DescribeAssetResponse (Maybe Text)

The ID of the PackagingGroup for the Asset.

describeAssetResponse_sourceArn :: Lens' DescribeAssetResponse (Maybe Text)

ARN of the source object in S3.

describeAssetResponse_sourceRoleArn :: Lens' DescribeAssetResponse (Maybe Text)

The IAM role_arn used to access the source S3 bucket.

describeAssetResponse_id :: Lens' DescribeAssetResponse (Maybe Text)

The unique identifier for the Asset.

describeAssetResponse_egressEndpoints :: Lens' DescribeAssetResponse (Maybe [EgressEndpoint])

The list of egress endpoints available for the Asset.

describeAssetResponse_tags :: Lens' DescribeAssetResponse (Maybe (HashMap Text Text))

Undocumented member.

describeAssetResponse_httpStatus :: Lens' DescribeAssetResponse Int

The response's http status code.

DeletePackagingConfiguration

deletePackagingConfiguration_id :: Lens' DeletePackagingConfiguration Text

The ID of the MediaPackage VOD PackagingConfiguration resource to delete.

deletePackagingConfigurationResponse_httpStatus :: Lens' DeletePackagingConfigurationResponse Int

The response's http status code.

ListPackagingGroups

listPackagingGroups_nextToken :: Lens' ListPackagingGroups (Maybe Text)

A token used to resume pagination from the end of a previous request.

listPackagingGroups_maxResults :: Lens' ListPackagingGroups (Maybe Natural)

Upper bound on number of records to return.

listPackagingGroupsResponse_packagingGroups :: Lens' ListPackagingGroupsResponse (Maybe [PackagingGroup])

A list of MediaPackage VOD PackagingGroup resources.

listPackagingGroupsResponse_nextToken :: Lens' ListPackagingGroupsResponse (Maybe Text)

A token that can be used to resume pagination from the end of the collection.

listPackagingGroupsResponse_httpStatus :: Lens' ListPackagingGroupsResponse Int

The response's http status code.

DeleteAsset

deleteAsset_id :: Lens' DeleteAsset Text

The ID of the MediaPackage VOD Asset resource to delete.

deleteAssetResponse_httpStatus :: Lens' DeleteAssetResponse Int

The response's http status code.

UpdatePackagingGroup

updatePackagingGroup_authorization :: Lens' UpdatePackagingGroup (Maybe Authorization)

Undocumented member.

updatePackagingGroup_id :: Lens' UpdatePackagingGroup Text

The ID of a MediaPackage VOD PackagingGroup resource.

updatePackagingGroupResponse_arn :: Lens' UpdatePackagingGroupResponse (Maybe Text)

The ARN of the PackagingGroup.

updatePackagingGroupResponse_authorization :: Lens' UpdatePackagingGroupResponse (Maybe Authorization)

Undocumented member.

updatePackagingGroupResponse_domainName :: Lens' UpdatePackagingGroupResponse (Maybe Text)

The fully qualified domain name for Assets in the PackagingGroup.

updatePackagingGroupResponse_id :: Lens' UpdatePackagingGroupResponse (Maybe Text)

The ID of the PackagingGroup.

updatePackagingGroupResponse_egressAccessLogs :: Lens' UpdatePackagingGroupResponse (Maybe EgressAccessLogs)

Undocumented member.

updatePackagingGroupResponse_tags :: Lens' UpdatePackagingGroupResponse (Maybe (HashMap Text Text))

Undocumented member.

updatePackagingGroupResponse_httpStatus :: Lens' UpdatePackagingGroupResponse Int

The response's http status code.

DeletePackagingGroup

deletePackagingGroup_id :: Lens' DeletePackagingGroup Text

The ID of the MediaPackage VOD PackagingGroup resource to delete.

deletePackagingGroupResponse_httpStatus :: Lens' DeletePackagingGroupResponse Int

The response's http status code.

CreateAsset

createAsset_resourceId :: Lens' CreateAsset (Maybe Text)

The resource ID to include in SPEKE key requests.

createAsset_tags :: Lens' CreateAsset (Maybe (HashMap Text Text))

Undocumented member.

createAsset_sourceArn :: Lens' CreateAsset Text

ARN of the source object in S3.

createAsset_id :: Lens' CreateAsset Text

The unique identifier for the Asset.

createAsset_packagingGroupId :: Lens' CreateAsset Text

The ID of the PackagingGroup for the Asset.

createAsset_sourceRoleArn :: Lens' CreateAsset Text

The IAM role ARN used to access the source S3 bucket.

createAssetResponse_resourceId :: Lens' CreateAssetResponse (Maybe Text)

The resource ID to include in SPEKE key requests.

createAssetResponse_arn :: Lens' CreateAssetResponse (Maybe Text)

The ARN of the Asset.

createAssetResponse_createdAt :: Lens' CreateAssetResponse (Maybe Text)

The time the Asset was initially submitted for Ingest.

createAssetResponse_packagingGroupId :: Lens' CreateAssetResponse (Maybe Text)

The ID of the PackagingGroup for the Asset.

createAssetResponse_sourceArn :: Lens' CreateAssetResponse (Maybe Text)

ARN of the source object in S3.

createAssetResponse_sourceRoleArn :: Lens' CreateAssetResponse (Maybe Text)

The IAM role_arn used to access the source S3 bucket.

createAssetResponse_id :: Lens' CreateAssetResponse (Maybe Text)

The unique identifier for the Asset.

createAssetResponse_egressEndpoints :: Lens' CreateAssetResponse (Maybe [EgressEndpoint])

The list of egress endpoints available for the Asset.

createAssetResponse_tags :: Lens' CreateAssetResponse (Maybe (HashMap Text Text))

Undocumented member.

createAssetResponse_httpStatus :: Lens' CreateAssetResponse Int

The response's http status code.

DescribePackagingConfiguration

describePackagingConfiguration_id :: Lens' DescribePackagingConfiguration Text

The ID of a MediaPackage VOD PackagingConfiguration resource.

describePackagingConfigurationResponse_hlsPackage :: Lens' DescribePackagingConfigurationResponse (Maybe HlsPackage)

Undocumented member.

describePackagingConfigurationResponse_arn :: Lens' DescribePackagingConfigurationResponse (Maybe Text)

The ARN of the PackagingConfiguration.

describePackagingConfigurationResponse_packagingGroupId :: Lens' DescribePackagingConfigurationResponse (Maybe Text)

The ID of a PackagingGroup.

describePackagingConfigurationResponse_dashPackage :: Lens' DescribePackagingConfigurationResponse (Maybe DashPackage)

Undocumented member.

describePackagingConfigurationResponse_mssPackage :: Lens' DescribePackagingConfigurationResponse (Maybe MssPackage)

Undocumented member.

describePackagingConfigurationResponse_id :: Lens' DescribePackagingConfigurationResponse (Maybe Text)

The ID of the PackagingConfiguration.

describePackagingConfigurationResponse_cmafPackage :: Lens' DescribePackagingConfigurationResponse (Maybe CmafPackage)

Undocumented member.

describePackagingConfigurationResponse_tags :: Lens' DescribePackagingConfigurationResponse (Maybe (HashMap Text Text))

Undocumented member.

describePackagingConfigurationResponse_httpStatus :: Lens' DescribePackagingConfigurationResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) for the resource. You can get this from the response to any request to the resource.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

A collection of tags associated with a resource

UntagResource

untagResource_tagKeys :: Lens' UntagResource [Text]

A comma-separated list of the tag keys to remove from the resource.

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) for the resource. You can get this from the response to any request to the resource.

CreatePackagingConfiguration

createPackagingConfiguration_hlsPackage :: Lens' CreatePackagingConfiguration (Maybe HlsPackage)

Undocumented member.

createPackagingConfiguration_dashPackage :: Lens' CreatePackagingConfiguration (Maybe DashPackage)

Undocumented member.

createPackagingConfiguration_mssPackage :: Lens' CreatePackagingConfiguration (Maybe MssPackage)

Undocumented member.

createPackagingConfiguration_cmafPackage :: Lens' CreatePackagingConfiguration (Maybe CmafPackage)

Undocumented member.

createPackagingConfiguration_tags :: Lens' CreatePackagingConfiguration (Maybe (HashMap Text Text))

Undocumented member.

createPackagingConfiguration_id :: Lens' CreatePackagingConfiguration Text

The ID of the PackagingConfiguration.

createPackagingConfiguration_packagingGroupId :: Lens' CreatePackagingConfiguration Text

The ID of a PackagingGroup.

createPackagingConfigurationResponse_hlsPackage :: Lens' CreatePackagingConfigurationResponse (Maybe HlsPackage)

Undocumented member.

createPackagingConfigurationResponse_arn :: Lens' CreatePackagingConfigurationResponse (Maybe Text)

The ARN of the PackagingConfiguration.

createPackagingConfigurationResponse_packagingGroupId :: Lens' CreatePackagingConfigurationResponse (Maybe Text)

The ID of a PackagingGroup.

createPackagingConfigurationResponse_dashPackage :: Lens' CreatePackagingConfigurationResponse (Maybe DashPackage)

Undocumented member.

createPackagingConfigurationResponse_mssPackage :: Lens' CreatePackagingConfigurationResponse (Maybe MssPackage)

Undocumented member.

createPackagingConfigurationResponse_id :: Lens' CreatePackagingConfigurationResponse (Maybe Text)

The ID of the PackagingConfiguration.

createPackagingConfigurationResponse_cmafPackage :: Lens' CreatePackagingConfigurationResponse (Maybe CmafPackage)

Undocumented member.

createPackagingConfigurationResponse_tags :: Lens' CreatePackagingConfigurationResponse (Maybe (HashMap Text Text))

Undocumented member.

createPackagingConfigurationResponse_httpStatus :: Lens' CreatePackagingConfigurationResponse Int

The response's http status code.

ListPackagingConfigurations

listPackagingConfigurations_packagingGroupId :: Lens' ListPackagingConfigurations (Maybe Text)

Returns MediaPackage VOD PackagingConfigurations associated with the specified PackagingGroup.

listPackagingConfigurations_nextToken :: Lens' ListPackagingConfigurations (Maybe Text)

A token used to resume pagination from the end of a previous request.

listPackagingConfigurations_maxResults :: Lens' ListPackagingConfigurations (Maybe Natural)

Upper bound on number of records to return.

listPackagingConfigurationsResponse_packagingConfigurations :: Lens' ListPackagingConfigurationsResponse (Maybe [PackagingConfiguration])

A list of MediaPackage VOD PackagingConfiguration resources.

listPackagingConfigurationsResponse_nextToken :: Lens' ListPackagingConfigurationsResponse (Maybe Text)

A token that can be used to resume pagination from the end of the collection.

listPackagingConfigurationsResponse_httpStatus :: Lens' ListPackagingConfigurationsResponse Int

The response's http status code.

ListAssets

listAssets_packagingGroupId :: Lens' ListAssets (Maybe Text)

Returns Assets associated with the specified PackagingGroup.

listAssets_nextToken :: Lens' ListAssets (Maybe Text)

A token used to resume pagination from the end of a previous request.

listAssets_maxResults :: Lens' ListAssets (Maybe Natural)

Upper bound on number of records to return.

listAssetsResponse_nextToken :: Lens' ListAssetsResponse (Maybe Text)

A token that can be used to resume pagination from the end of the collection.

listAssetsResponse_assets :: Lens' ListAssetsResponse (Maybe [AssetShallow])

A list of MediaPackage VOD Asset resources.

listAssetsResponse_httpStatus :: Lens' ListAssetsResponse Int

The response's http status code.

Types

AssetShallow

assetShallow_resourceId :: Lens' AssetShallow (Maybe Text)

The resource ID to include in SPEKE key requests.

assetShallow_arn :: Lens' AssetShallow (Maybe Text)

The ARN of the Asset.

assetShallow_createdAt :: Lens' AssetShallow (Maybe Text)

The time the Asset was initially submitted for Ingest.

assetShallow_packagingGroupId :: Lens' AssetShallow (Maybe Text)

The ID of the PackagingGroup for the Asset.

assetShallow_sourceArn :: Lens' AssetShallow (Maybe Text)

ARN of the source object in S3.

assetShallow_sourceRoleArn :: Lens' AssetShallow (Maybe Text)

The IAM role ARN used to access the source S3 bucket.

assetShallow_id :: Lens' AssetShallow (Maybe Text)

The unique identifier for the Asset.

assetShallow_tags :: Lens' AssetShallow (Maybe (HashMap Text Text))

Undocumented member.

Authorization

authorization_secretsRoleArn :: Lens' Authorization Text

The Amazon Resource Name (ARN) for the IAM role that allows MediaPackage to communicate with AWS Secrets Manager.

authorization_cdnIdentifierSecret :: Lens' Authorization Text

The Amazon Resource Name (ARN) for the secret in AWS Secrets Manager that is used for CDN authorization.

CmafEncryption

cmafEncryption_constantInitializationVector :: Lens' CmafEncryption (Maybe Text)

An optional 128-bit, 16-byte hex value represented by a 32-character string, used in conjunction with the key for encrypting blocks. If you don't specify a value, then MediaPackage creates the constant initialization vector (IV).

cmafEncryption_spekeKeyProvider :: Lens' CmafEncryption SpekeKeyProvider

Undocumented member.

CmafPackage

cmafPackage_includeEncoderConfigurationInSegments :: Lens' CmafPackage (Maybe Bool)

When includeEncoderConfigurationInSegments is set to true, MediaPackage places your encoder's Sequence Parameter Set (SPS), Picture Parameter Set (PPS), and Video Parameter Set (VPS) metadata in every video segment instead of in the init fragment. This lets you use different SPS/PPS/VPS settings for your assets during content playback.

cmafPackage_segmentDurationSeconds :: Lens' CmafPackage (Maybe Int)

Duration (in seconds) of each fragment. Actual fragments will be rounded to the nearest multiple of the source fragment duration.

cmafPackage_encryption :: Lens' CmafPackage (Maybe CmafEncryption)

Undocumented member.

cmafPackage_hlsManifests :: Lens' CmafPackage [HlsManifest]

A list of HLS manifest configurations.

DashEncryption

dashEncryption_spekeKeyProvider :: Lens' DashEncryption SpekeKeyProvider

Undocumented member.

DashManifest

dashManifest_minBufferTimeSeconds :: Lens' DashManifest (Maybe Int)

Minimum duration (in seconds) that a player will buffer media before starting the presentation.

dashManifest_manifestName :: Lens' DashManifest (Maybe Text)

An optional string to include in the name of the manifest.

dashManifest_profile :: Lens' DashManifest (Maybe Profile)

The Dynamic Adaptive Streaming over HTTP (DASH) profile type. When set to "HBBTV_1_5", HbbTV 1.5 compliant output is enabled.

dashManifest_streamSelection :: Lens' DashManifest (Maybe StreamSelection)

Undocumented member.

dashManifest_manifestLayout :: Lens' DashManifest (Maybe ManifestLayout)

Determines the position of some tags in the Media Presentation Description (MPD). When set to FULL, elements like SegmentTemplate and ContentProtection are included in each Representation. When set to COMPACT, duplicate elements are combined and presented at the AdaptationSet level.

DashPackage

dashPackage_includeEncoderConfigurationInSegments :: Lens' DashPackage (Maybe Bool)

When includeEncoderConfigurationInSegments is set to true, MediaPackage places your encoder's Sequence Parameter Set (SPS), Picture Parameter Set (PPS), and Video Parameter Set (VPS) metadata in every video segment instead of in the init fragment. This lets you use different SPS/PPS/VPS settings for your assets during content playback.

dashPackage_segmentTemplateFormat :: Lens' DashPackage (Maybe SegmentTemplateFormat)

Determines the type of SegmentTemplate included in the Media Presentation Description (MPD). When set to NUMBER_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Number$ media URLs. When set to TIME_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Time$ media URLs. When set to NUMBER_WITH_DURATION, only a duration is included in each SegmentTemplate, with $Number$ media URLs.

dashPackage_segmentDurationSeconds :: Lens' DashPackage (Maybe Int)

Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source segment duration.

dashPackage_encryption :: Lens' DashPackage (Maybe DashEncryption)

Undocumented member.

dashPackage_periodTriggers :: Lens' DashPackage (Maybe [PeriodTriggersElement])

A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation Description (MPD) will be partitioned into multiple periods. If empty, the content will not be partitioned into more than one period. If the list contains "ADS", new periods will be created where the Asset contains SCTE-35 ad markers.

dashPackage_dashManifests :: Lens' DashPackage [DashManifest]

A list of DASH manifest configurations.

EgressAccessLogs

egressAccessLogs_logGroupName :: Lens' EgressAccessLogs (Maybe Text)

Customize the log group name.

EgressEndpoint

egressEndpoint_status :: Lens' EgressEndpoint (Maybe Text)

The current processing status of the asset used for the packaging configuration. The status can be either QUEUED, PROCESSING, PLAYABLE, or FAILED. Status information won't be available for most assets ingested before 2021-09-30.

egressEndpoint_url :: Lens' EgressEndpoint (Maybe Text)

The URL of the parent manifest for the repackaged Asset.

egressEndpoint_packagingConfigurationId :: Lens' EgressEndpoint (Maybe Text)

The ID of the PackagingConfiguration being applied to the Asset.

HlsEncryption

hlsEncryption_encryptionMethod :: Lens' HlsEncryption (Maybe EncryptionMethod)

The encryption method to use.

hlsEncryption_constantInitializationVector :: Lens' HlsEncryption (Maybe Text)

A constant initialization vector for encryption (optional). When not specified the initialization vector will be periodically rotated.

hlsEncryption_spekeKeyProvider :: Lens' HlsEncryption SpekeKeyProvider

Undocumented member.

HlsManifest

hlsManifest_manifestName :: Lens' HlsManifest (Maybe Text)

An optional string to include in the name of the manifest.

hlsManifest_programDateTimeIntervalSeconds :: Lens' HlsManifest (Maybe Int)

The interval (in seconds) between each EXT-X-PROGRAM-DATE-TIME tag inserted into manifests. Additionally, when an interval is specified ID3Timed Metadata messages will be generated every 5 seconds using the ingest time of the content. If the interval is not specified, or set to 0, then no EXT-X-PROGRAM-DATE-TIME tags will be inserted into manifests and no ID3Timed Metadata messages will be generated. Note that irrespective of this parameter, if any ID3 Timed Metadata is found in HTTP Live Streaming (HLS) input, it will be passed through to HLS output.

hlsManifest_streamSelection :: Lens' HlsManifest (Maybe StreamSelection)

Undocumented member.

hlsManifest_adMarkers :: Lens' HlsManifest (Maybe AdMarkers)

This setting controls how ad markers are included in the packaged OriginEndpoint. "NONE" will omit all SCTE-35 ad markers from the output. "PASSTHROUGH" causes the manifest to contain a copy of the SCTE-35 ad markers (comments) taken directly from the input HTTP Live Streaming (HLS) manifest. "SCTE35_ENHANCED" generates ad markers and blackout tags based on SCTE-35 messages in the input source.

hlsManifest_includeIframeOnlyStream :: Lens' HlsManifest (Maybe Bool)

When enabled, an I-Frame only stream will be included in the output.

hlsManifest_repeatExtXKey :: Lens' HlsManifest (Maybe Bool)

When enabled, the EXT-X-KEY tag will be repeated in output manifests.

HlsPackage

hlsPackage_useAudioRenditionGroup :: Lens' HlsPackage (Maybe Bool)

When enabled, audio streams will be placed in rendition groups in the output.

hlsPackage_includeDvbSubtitles :: Lens' HlsPackage (Maybe Bool)

When enabled, MediaPackage passes through digital video broadcasting (DVB) subtitles into the output.

hlsPackage_segmentDurationSeconds :: Lens' HlsPackage (Maybe Int)

Duration (in seconds) of each fragment. Actual fragments will be rounded to the nearest multiple of the source fragment duration.

hlsPackage_encryption :: Lens' HlsPackage (Maybe HlsEncryption)

Undocumented member.

hlsPackage_hlsManifests :: Lens' HlsPackage [HlsManifest]

A list of HLS manifest configurations.

MssEncryption

mssEncryption_spekeKeyProvider :: Lens' MssEncryption SpekeKeyProvider

Undocumented member.

MssManifest

mssManifest_manifestName :: Lens' MssManifest (Maybe Text)

An optional string to include in the name of the manifest.

mssManifest_streamSelection :: Lens' MssManifest (Maybe StreamSelection)

Undocumented member.

MssPackage

mssPackage_segmentDurationSeconds :: Lens' MssPackage (Maybe Int)

The duration (in seconds) of each segment.

mssPackage_encryption :: Lens' MssPackage (Maybe MssEncryption)

Undocumented member.

mssPackage_mssManifests :: Lens' MssPackage [MssManifest]

A list of MSS manifest configurations.

PackagingConfiguration

packagingConfiguration_hlsPackage :: Lens' PackagingConfiguration (Maybe HlsPackage)

Undocumented member.

packagingConfiguration_arn :: Lens' PackagingConfiguration (Maybe Text)

The ARN of the PackagingConfiguration.

packagingConfiguration_packagingGroupId :: Lens' PackagingConfiguration (Maybe Text)

The ID of a PackagingGroup.

packagingConfiguration_dashPackage :: Lens' PackagingConfiguration (Maybe DashPackage)

Undocumented member.

packagingConfiguration_mssPackage :: Lens' PackagingConfiguration (Maybe MssPackage)

Undocumented member.

packagingConfiguration_id :: Lens' PackagingConfiguration (Maybe Text)

The ID of the PackagingConfiguration.

packagingConfiguration_cmafPackage :: Lens' PackagingConfiguration (Maybe CmafPackage)

Undocumented member.

packagingConfiguration_tags :: Lens' PackagingConfiguration (Maybe (HashMap Text Text))

Undocumented member.

PackagingGroup

packagingGroup_arn :: Lens' PackagingGroup (Maybe Text)

The ARN of the PackagingGroup.

packagingGroup_authorization :: Lens' PackagingGroup (Maybe Authorization)

Undocumented member.

packagingGroup_domainName :: Lens' PackagingGroup (Maybe Text)

The fully qualified domain name for Assets in the PackagingGroup.

packagingGroup_id :: Lens' PackagingGroup (Maybe Text)

The ID of the PackagingGroup.

packagingGroup_egressAccessLogs :: Lens' PackagingGroup (Maybe EgressAccessLogs)

Undocumented member.

packagingGroup_tags :: Lens' PackagingGroup (Maybe (HashMap Text Text))

Undocumented member.

SpekeKeyProvider

spekeKeyProvider_systemIds :: Lens' SpekeKeyProvider [Text]

The system IDs to include in key requests.

spekeKeyProvider_url :: Lens' SpekeKeyProvider Text

The URL of the external key provider service.

spekeKeyProvider_roleArn :: Lens' SpekeKeyProvider Text

An Amazon Resource Name (ARN) of an IAM role that AWS Elemental MediaPackage will assume when accessing the key provider service.

StreamSelection

streamSelection_streamOrder :: Lens' StreamSelection (Maybe StreamOrder)

A directive that determines the order of streams in the output.

streamSelection_minVideoBitsPerSecond :: Lens' StreamSelection (Maybe Int)

The minimum video bitrate (bps) to include in output.

streamSelection_maxVideoBitsPerSecond :: Lens' StreamSelection (Maybe Int)

The maximum video bitrate (bps) to include in output.