Network.AWS.MediaPackageVOD.ListPackagingGroups

Description

Returns a collection of MediaPackage VOD PackagingGroup resources.

This operation returns paginated results.

Creating a Request

data ListPackagingGroups

See: newListPackagingGroups smart constructor.

Constructors

ListPackagingGroups'

Fields

nextToken :: Maybe Text

A token used to resume pagination from the end of a previous request.

maxResults :: Maybe Natural

Upper bound on number of records to return.

newListPackagingGroups :: ListPackagingGroups

Create a value of ListPackagingGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListPackagingGroups', listPackagingGroups_nextToken - A token used to resume pagination from the end of a previous request.

$sel:maxResults:ListPackagingGroups', listPackagingGroups_maxResults - Upper bound on number of records to return.

Request Lenses

listPackagingGroups_nextToken :: Lens' ListPackagingGroups (Maybe Text)

A token used to resume pagination from the end of a previous request.

listPackagingGroups_maxResults :: Lens' ListPackagingGroups (Maybe Natural)

Upper bound on number of records to return.

Destructuring the Response

data ListPackagingGroupsResponse

See: newListPackagingGroupsResponse smart constructor.

Constructors

ListPackagingGroupsResponse'

Fields

packagingGroups :: Maybe [PackagingGroup]

A list of MediaPackage VOD PackagingGroup resources.

nextToken :: Maybe Text

A token that can be used to resume pagination from the end of the collection.

httpStatus :: Int

The response's http status code.

newListPackagingGroupsResponse

Arguments

:: Int
$sel:httpStatus:ListPackagingGroupsResponse'

-> ListPackagingGroupsResponse

Create a value of ListPackagingGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packagingGroups:ListPackagingGroupsResponse', listPackagingGroupsResponse_packagingGroups - A list of MediaPackage VOD PackagingGroup resources.

$sel:nextToken:ListPackagingGroups', listPackagingGroupsResponse_nextToken - A token that can be used to resume pagination from the end of the collection.

$sel:httpStatus:ListPackagingGroupsResponse', listPackagingGroupsResponse_httpStatus - The response's http status code.

Response Lenses

listPackagingGroupsResponse_packagingGroups :: Lens' ListPackagingGroupsResponse (Maybe [PackagingGroup])

A list of MediaPackage VOD PackagingGroup resources.

listPackagingGroupsResponse_nextToken :: Lens' ListPackagingGroupsResponse (Maybe Text)

A token that can be used to resume pagination from the end of the collection.

listPackagingGroupsResponse_httpStatus :: Lens' ListPackagingGroupsResponse Int

The response's http status code.