Network.AWS.MediaPackageVOD.Types.AssetShallow

Description

Documentation

data AssetShallow

A MediaPackage VOD Asset resource.

See: newAssetShallow smart constructor.

Constructors

AssetShallow'

Fields

resourceId :: Maybe Text

The resource ID to include in SPEKE key requests.

arn :: Maybe Text

The ARN of the Asset.

createdAt :: Maybe Text

The time the Asset was initially submitted for Ingest.

packagingGroupId :: Maybe Text

The ID of the PackagingGroup for the Asset.

sourceArn :: Maybe Text

ARN of the source object in S3.

sourceRoleArn :: Maybe Text

The IAM role ARN used to access the source S3 bucket.

id :: Maybe Text

The unique identifier for the Asset.

tags :: Maybe (HashMap Text Text)

newAssetShallow :: AssetShallow

Create a value of AssetShallow with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceId:AssetShallow', assetShallow_resourceId - The resource ID to include in SPEKE key requests.

$sel:arn:AssetShallow', assetShallow_arn - The ARN of the Asset.

$sel:createdAt:AssetShallow', assetShallow_createdAt - The time the Asset was initially submitted for Ingest.

$sel:packagingGroupId:AssetShallow', assetShallow_packagingGroupId - The ID of the PackagingGroup for the Asset.

$sel:sourceArn:AssetShallow', assetShallow_sourceArn - ARN of the source object in S3.

$sel:sourceRoleArn:AssetShallow', assetShallow_sourceRoleArn - The IAM role ARN used to access the source S3 bucket.

$sel:id:AssetShallow', assetShallow_id - The unique identifier for the Asset.

$sel:tags:AssetShallow', assetShallow_tags - Undocumented member.

assetShallow_resourceId :: Lens' AssetShallow (Maybe Text)

The resource ID to include in SPEKE key requests.

assetShallow_arn :: Lens' AssetShallow (Maybe Text)

The ARN of the Asset.

assetShallow_createdAt :: Lens' AssetShallow (Maybe Text)

The time the Asset was initially submitted for Ingest.

assetShallow_packagingGroupId :: Lens' AssetShallow (Maybe Text)

The ID of the PackagingGroup for the Asset.

assetShallow_sourceArn :: Lens' AssetShallow (Maybe Text)

ARN of the source object in S3.

assetShallow_sourceRoleArn :: Lens' AssetShallow (Maybe Text)

The IAM role ARN used to access the source S3 bucket.

assetShallow_id :: Lens' AssetShallow (Maybe Text)

The unique identifier for the Asset.

assetShallow_tags :: Lens' AssetShallow (Maybe (HashMap Text Text))

Undocumented member.