Network.AWS.MediaPackageVOD.Types.DashManifest

Description

Documentation

data DashManifest

A DASH manifest configuration.

See: newDashManifest smart constructor.

Constructors

DashManifest'

Fields

minBufferTimeSeconds :: Maybe Int

Minimum duration (in seconds) that a player will buffer media before starting the presentation.

manifestName :: Maybe Text

An optional string to include in the name of the manifest.

profile :: Maybe Profile

The Dynamic Adaptive Streaming over HTTP (DASH) profile type. When set to "HBBTV_1_5", HbbTV 1.5 compliant output is enabled.

streamSelection :: Maybe StreamSelection

manifestLayout :: Maybe ManifestLayout

Determines the position of some tags in the Media Presentation Description (MPD). When set to FULL, elements like SegmentTemplate and ContentProtection are included in each Representation. When set to COMPACT, duplicate elements are combined and presented at the AdaptationSet level.

newDashManifest :: DashManifest

Create a value of DashManifest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:minBufferTimeSeconds:DashManifest', dashManifest_minBufferTimeSeconds - Minimum duration (in seconds) that a player will buffer media before starting the presentation.

$sel:manifestName:DashManifest', dashManifest_manifestName - An optional string to include in the name of the manifest.

$sel:profile:DashManifest', dashManifest_profile - The Dynamic Adaptive Streaming over HTTP (DASH) profile type. When set to "HBBTV_1_5", HbbTV 1.5 compliant output is enabled.

$sel:streamSelection:DashManifest', dashManifest_streamSelection - Undocumented member.

$sel:manifestLayout:DashManifest', dashManifest_manifestLayout - Determines the position of some tags in the Media Presentation Description (MPD). When set to FULL, elements like SegmentTemplate and ContentProtection are included in each Representation. When set to COMPACT, duplicate elements are combined and presented at the AdaptationSet level.

dashManifest_minBufferTimeSeconds :: Lens' DashManifest (Maybe Int)

Minimum duration (in seconds) that a player will buffer media before starting the presentation.

dashManifest_manifestName :: Lens' DashManifest (Maybe Text)

An optional string to include in the name of the manifest.

dashManifest_profile :: Lens' DashManifest (Maybe Profile)

The Dynamic Adaptive Streaming over HTTP (DASH) profile type. When set to "HBBTV_1_5", HbbTV 1.5 compliant output is enabled.

dashManifest_streamSelection :: Lens' DashManifest (Maybe StreamSelection)

Undocumented member.

dashManifest_manifestLayout :: Lens' DashManifest (Maybe ManifestLayout)

Determines the position of some tags in the Media Presentation Description (MPD). When set to FULL, elements like SegmentTemplate and ContentProtection are included in each Representation. When set to COMPACT, duplicate elements are combined and presented at the AdaptationSet level.