Network.AWS.MediaPackageVOD.Types.DashPackage

Description

Documentation

data DashPackage

A Dynamic Adaptive Streaming over HTTP (DASH) packaging configuration.

See: newDashPackage smart constructor.

Constructors

DashPackage'

Fields

includeEncoderConfigurationInSegments :: Maybe Bool

When includeEncoderConfigurationInSegments is set to true, MediaPackage places your encoder's Sequence Parameter Set (SPS), Picture Parameter Set (PPS), and Video Parameter Set (VPS) metadata in every video segment instead of in the init fragment. This lets you use different SPS/PPS/VPS settings for your assets during content playback.

segmentTemplateFormat :: Maybe SegmentTemplateFormat

Determines the type of SegmentTemplate included in the Media Presentation Description (MPD). When set to NUMBER_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Number$ media URLs. When set to TIME_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Time$ media URLs. When set to NUMBER_WITH_DURATION, only a duration is included in each SegmentTemplate, with $Number$ media URLs.

segmentDurationSeconds :: Maybe Int

Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source segment duration.

encryption :: Maybe DashEncryption

periodTriggers :: Maybe [PeriodTriggersElement]

A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation Description (MPD) will be partitioned into multiple periods. If empty, the content will not be partitioned into more than one period. If the list contains "ADS", new periods will be created where the Asset contains SCTE-35 ad markers.

dashManifests :: [DashManifest]

A list of DASH manifest configurations.

newDashPackage :: DashPackage

Create a value of DashPackage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:includeEncoderConfigurationInSegments:DashPackage', dashPackage_includeEncoderConfigurationInSegments - When includeEncoderConfigurationInSegments is set to true, MediaPackage places your encoder's Sequence Parameter Set (SPS), Picture Parameter Set (PPS), and Video Parameter Set (VPS) metadata in every video segment instead of in the init fragment. This lets you use different SPS/PPS/VPS settings for your assets during content playback.

$sel:segmentTemplateFormat:DashPackage', dashPackage_segmentTemplateFormat - Determines the type of SegmentTemplate included in the Media Presentation Description (MPD). When set to NUMBER_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Number$ media URLs. When set to TIME_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Time$ media URLs. When set to NUMBER_WITH_DURATION, only a duration is included in each SegmentTemplate, with $Number$ media URLs.

$sel:segmentDurationSeconds:DashPackage', dashPackage_segmentDurationSeconds - Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source segment duration.

$sel:encryption:DashPackage', dashPackage_encryption - Undocumented member.

$sel:periodTriggers:DashPackage', dashPackage_periodTriggers - A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation Description (MPD) will be partitioned into multiple periods. If empty, the content will not be partitioned into more than one period. If the list contains "ADS", new periods will be created where the Asset contains SCTE-35 ad markers.

$sel:dashManifests:DashPackage', dashPackage_dashManifests - A list of DASH manifest configurations.

dashPackage_includeEncoderConfigurationInSegments :: Lens' DashPackage (Maybe Bool)

When includeEncoderConfigurationInSegments is set to true, MediaPackage places your encoder's Sequence Parameter Set (SPS), Picture Parameter Set (PPS), and Video Parameter Set (VPS) metadata in every video segment instead of in the init fragment. This lets you use different SPS/PPS/VPS settings for your assets during content playback.

dashPackage_segmentTemplateFormat :: Lens' DashPackage (Maybe SegmentTemplateFormat)

Determines the type of SegmentTemplate included in the Media Presentation Description (MPD). When set to NUMBER_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Number$ media URLs. When set to TIME_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Time$ media URLs. When set to NUMBER_WITH_DURATION, only a duration is included in each SegmentTemplate, with $Number$ media URLs.

dashPackage_segmentDurationSeconds :: Lens' DashPackage (Maybe Int)

Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source segment duration.

dashPackage_encryption :: Lens' DashPackage (Maybe DashEncryption)

Undocumented member.

dashPackage_periodTriggers :: Lens' DashPackage (Maybe [PeriodTriggersElement])

A list of triggers that controls when the outgoing Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation Description (MPD) will be partitioned into multiple periods. If empty, the content will not be partitioned into more than one period. If the list contains "ADS", new periods will be created where the Asset contains SCTE-35 ad markers.

dashPackage_dashManifests :: Lens' DashPackage [DashManifest]

A list of DASH manifest configurations.