Network.AWS.MediaStore

Description

Derived from API version 2017-09-01 of the AWS service descriptions, licensed under Apache 2.0.

An AWS Elemental MediaStore container is a namespace that holds folders and objects. You use a container endpoint to create, read, and delete objects.

Service Configuration

defaultService :: Service

API version 2017-09-01 of the Amazon Elemental MediaStore SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by MediaStore.

ContainerInUseException

_ContainerInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The container that you specified in the request already exists or is being updated.

InternalServerError

_InternalServerError :: AsError a => Getting (First ServiceError) a ServiceError

The service is temporarily unavailable.

PolicyNotFoundException

_PolicyNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The policy that you specified in the request does not exist.

CorsPolicyNotFoundException

_CorsPolicyNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The CORS policy that you specified in the request does not exist.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

A service limit has been exceeded.

ContainerNotFoundException

_ContainerNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The container that you specified in the request does not exist.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

PutLifecyclePolicy

data PutLifecyclePolicy

See: newPutLifecyclePolicy smart constructor.

Constructors

PutLifecyclePolicy' Text Text

newPutLifecyclePolicy

Arguments

:: Text
$sel:containerName:PutLifecyclePolicy'

-> Text
$sel:lifecyclePolicy:PutLifecyclePolicy'

-> PutLifecyclePolicy

Create a value of PutLifecyclePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerName:PutLifecyclePolicy', putLifecyclePolicy_containerName - The name of the container that you want to assign the object lifecycle policy to.

$sel:lifecyclePolicy:PutLifecyclePolicy', putLifecyclePolicy_lifecyclePolicy - The object lifecycle policy to apply to the container.

data PutLifecyclePolicyResponse

See: newPutLifecyclePolicyResponse smart constructor.

Constructors

PutLifecyclePolicyResponse' Int

newPutLifecyclePolicyResponse

Arguments

:: Int
$sel:httpStatus:PutLifecyclePolicyResponse'

-> PutLifecyclePolicyResponse

Create a value of PutLifecyclePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutLifecyclePolicyResponse', putLifecyclePolicyResponse_httpStatus - The response's http status code.

PutCorsPolicy

data PutCorsPolicy

See: newPutCorsPolicy smart constructor.

Constructors

PutCorsPolicy' Text (NonEmpty CorsRule)

newPutCorsPolicy

Arguments

:: Text
$sel:containerName:PutCorsPolicy'

-> NonEmpty CorsRule
$sel:corsPolicy:PutCorsPolicy'

-> PutCorsPolicy

Create a value of PutCorsPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerName:PutCorsPolicy', putCorsPolicy_containerName - The name of the container that you want to assign the CORS policy to.

$sel:corsPolicy:PutCorsPolicy', putCorsPolicy_corsPolicy - The CORS policy to apply to the container.

data PutCorsPolicyResponse

See: newPutCorsPolicyResponse smart constructor.

Constructors

PutCorsPolicyResponse' Int

newPutCorsPolicyResponse

Arguments

:: Int
$sel:httpStatus:PutCorsPolicyResponse'

-> PutCorsPolicyResponse

Create a value of PutCorsPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutCorsPolicyResponse', putCorsPolicyResponse_httpStatus - The response's http status code.

DeleteContainer

data DeleteContainer

See: newDeleteContainer smart constructor.

Constructors

DeleteContainer' Text

newDeleteContainer

Arguments

:: Text
$sel:containerName:DeleteContainer'

-> DeleteContainer

Create a value of DeleteContainer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerName:DeleteContainer', deleteContainer_containerName - The name of the container to delete.

data DeleteContainerResponse

See: newDeleteContainerResponse smart constructor.

Constructors

DeleteContainerResponse' Int

newDeleteContainerResponse

Arguments

:: Int
$sel:httpStatus:DeleteContainerResponse'

-> DeleteContainerResponse

Create a value of DeleteContainerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteContainerResponse', deleteContainerResponse_httpStatus - The response's http status code.

PutContainerPolicy

data PutContainerPolicy

See: newPutContainerPolicy smart constructor.

Constructors

PutContainerPolicy' Text Text

newPutContainerPolicy

Arguments

:: Text
$sel:containerName:PutContainerPolicy'

-> Text
$sel:policy:PutContainerPolicy'

-> PutContainerPolicy

Create a value of PutContainerPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerName:PutContainerPolicy', putContainerPolicy_containerName - The name of the container.

$sel:policy:PutContainerPolicy', putContainerPolicy_policy - The contents of the policy, which includes the following:

One Version tag

One Statement tag that contains the standard tags for the policy.

data PutContainerPolicyResponse

See: newPutContainerPolicyResponse smart constructor.

Constructors

PutContainerPolicyResponse' Int

newPutContainerPolicyResponse

Arguments

:: Int
$sel:httpStatus:PutContainerPolicyResponse'

-> PutContainerPolicyResponse

Create a value of PutContainerPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutContainerPolicyResponse', putContainerPolicyResponse_httpStatus - The response's http status code.

GetCorsPolicy

data GetCorsPolicy

See: newGetCorsPolicy smart constructor.

Constructors

GetCorsPolicy' Text

newGetCorsPolicy

Arguments

:: Text
$sel:containerName:GetCorsPolicy'

-> GetCorsPolicy

Create a value of GetCorsPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerName:GetCorsPolicy', getCorsPolicy_containerName - The name of the container that the policy is assigned to.

data GetCorsPolicyResponse

See: newGetCorsPolicyResponse smart constructor.

Constructors

GetCorsPolicyResponse' Int (NonEmpty CorsRule)

newGetCorsPolicyResponse

Arguments

:: Int
$sel:httpStatus:GetCorsPolicyResponse'

-> NonEmpty CorsRule
$sel:corsPolicy:GetCorsPolicyResponse'

-> GetCorsPolicyResponse

Create a value of GetCorsPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetCorsPolicyResponse', getCorsPolicyResponse_httpStatus - The response's http status code.

$sel:corsPolicy:GetCorsPolicyResponse', getCorsPolicyResponse_corsPolicy - The CORS policy assigned to the container.

UntagResource

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource' Text [Text]

newUntagResource

Arguments

:: Text
$sel:resource:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resource:UntagResource', untagResource_resource - The Amazon Resource Name (ARN) for the container.

$sel:tagKeys:UntagResource', untagResource_tagKeys - A comma-separated list of keys for tags that you want to remove from the container. For example, if your container has two tags (customer:CompanyA and priority:High) and you want to remove one of the tags (priority:High), you specify the key for the tag that you want to remove (priority).

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse' Int

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

TagResource

data TagResource

See: newTagResource smart constructor.

Constructors

TagResource' Text (NonEmpty Tag)

newTagResource

Arguments

:: Text
$sel:resource:TagResource'

-> NonEmpty Tag
$sel:tags:TagResource'

-> TagResource

Create a value of TagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resource:TagResource', tagResource_resource - The Amazon Resource Name (ARN) for the container.

$sel:tags:TagResource', tagResource_tags - An array of key:value pairs that you want to add to the container. You need to specify only the tags that you want to add or update. For example, suppose a container already has two tags (customer:CompanyA and priority:High). You want to change the priority tag and also add a third tag (type:Contract). For TagResource, you specify the following tags: priority:Medium, type:Contract. The result is that your container has three tags: customer:CompanyA, priority:Medium, and type:Contract.

data TagResourceResponse

See: newTagResourceResponse smart constructor.

Constructors

TagResourceResponse' Int

newTagResourceResponse

Arguments

:: Int
$sel:httpStatus:TagResourceResponse'

-> TagResourceResponse

Create a value of TagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:TagResourceResponse', tagResourceResponse_httpStatus - The response's http status code.

DeleteMetricPolicy

data DeleteMetricPolicy

See: newDeleteMetricPolicy smart constructor.

Constructors

DeleteMetricPolicy' Text

newDeleteMetricPolicy

Arguments

:: Text
$sel:containerName:DeleteMetricPolicy'

-> DeleteMetricPolicy

Create a value of DeleteMetricPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerName:DeleteMetricPolicy', deleteMetricPolicy_containerName - The name of the container that is associated with the metric policy that you want to delete.

data DeleteMetricPolicyResponse

See: newDeleteMetricPolicyResponse smart constructor.

Constructors

DeleteMetricPolicyResponse' Int

newDeleteMetricPolicyResponse

Arguments

:: Int
$sel:httpStatus:DeleteMetricPolicyResponse'

-> DeleteMetricPolicyResponse

Create a value of DeleteMetricPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteMetricPolicyResponse', deleteMetricPolicyResponse_httpStatus - The response's http status code.

DescribeContainer

data DescribeContainer

See: newDescribeContainer smart constructor.

Constructors

DescribeContainer' (Maybe Text)

newDescribeContainer :: DescribeContainer

Create a value of DescribeContainer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerName:DescribeContainer', describeContainer_containerName - The name of the container to query.

data DescribeContainerResponse

See: newDescribeContainerResponse smart constructor.

Constructors

DescribeContainerResponse' (Maybe Container) Int

newDescribeContainerResponse

Arguments

:: Int
$sel:httpStatus:DescribeContainerResponse'

-> DescribeContainerResponse

Create a value of DescribeContainerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:container:DescribeContainerResponse', describeContainerResponse_container - The name of the queried container.

$sel:httpStatus:DescribeContainerResponse', describeContainerResponse_httpStatus - The response's http status code.

GetMetricPolicy

data GetMetricPolicy

See: newGetMetricPolicy smart constructor.

Constructors

GetMetricPolicy' Text

newGetMetricPolicy

Arguments

:: Text
$sel:containerName:GetMetricPolicy'

-> GetMetricPolicy

Create a value of GetMetricPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerName:GetMetricPolicy', getMetricPolicy_containerName - The name of the container that is associated with the metric policy.

data GetMetricPolicyResponse

See: newGetMetricPolicyResponse smart constructor.

Constructors

GetMetricPolicyResponse' Int MetricPolicy

newGetMetricPolicyResponse

Arguments

:: Int
$sel:httpStatus:GetMetricPolicyResponse'

-> MetricPolicy
$sel:metricPolicy:GetMetricPolicyResponse'

-> GetMetricPolicyResponse

Create a value of GetMetricPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetMetricPolicyResponse', getMetricPolicyResponse_httpStatus - The response's http status code.

$sel:metricPolicy:GetMetricPolicyResponse', getMetricPolicyResponse_metricPolicy - The metric policy that is associated with the specific container.

StopAccessLogging

data StopAccessLogging

See: newStopAccessLogging smart constructor.

Constructors

StopAccessLogging' Text

newStopAccessLogging

Arguments

:: Text
$sel:containerName:StopAccessLogging'

-> StopAccessLogging

Create a value of StopAccessLogging with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerName:StopAccessLogging', stopAccessLogging_containerName - The name of the container that you want to stop access logging on.

data StopAccessLoggingResponse

See: newStopAccessLoggingResponse smart constructor.

Constructors

StopAccessLoggingResponse' Int

newStopAccessLoggingResponse

Arguments

:: Int
$sel:httpStatus:StopAccessLoggingResponse'

-> StopAccessLoggingResponse

Create a value of StopAccessLoggingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StopAccessLoggingResponse', stopAccessLoggingResponse_httpStatus - The response's http status code.

StartAccessLogging

data StartAccessLogging

See: newStartAccessLogging smart constructor.

Constructors

StartAccessLogging' Text

newStartAccessLogging

Arguments

:: Text
$sel:containerName:StartAccessLogging'

-> StartAccessLogging

Create a value of StartAccessLogging with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerName:StartAccessLogging', startAccessLogging_containerName - The name of the container that you want to start access logging on.

data StartAccessLoggingResponse

See: newStartAccessLoggingResponse smart constructor.

Constructors

StartAccessLoggingResponse' Int

newStartAccessLoggingResponse

Arguments

:: Int
$sel:httpStatus:StartAccessLoggingResponse'

-> StartAccessLoggingResponse

Create a value of StartAccessLoggingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:StartAccessLoggingResponse', startAccessLoggingResponse_httpStatus - The response's http status code.

DeleteLifecyclePolicy

data DeleteLifecyclePolicy

See: newDeleteLifecyclePolicy smart constructor.

Constructors

DeleteLifecyclePolicy' Text

newDeleteLifecyclePolicy

Arguments

:: Text
$sel:containerName:DeleteLifecyclePolicy'

-> DeleteLifecyclePolicy

Create a value of DeleteLifecyclePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerName:DeleteLifecyclePolicy', deleteLifecyclePolicy_containerName - The name of the container that holds the object lifecycle policy.

data DeleteLifecyclePolicyResponse

See: newDeleteLifecyclePolicyResponse smart constructor.

Constructors

DeleteLifecyclePolicyResponse' Int

newDeleteLifecyclePolicyResponse

Arguments

:: Int
$sel:httpStatus:DeleteLifecyclePolicyResponse'

-> DeleteLifecyclePolicyResponse

Create a value of DeleteLifecyclePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteLifecyclePolicyResponse', deleteLifecyclePolicyResponse_httpStatus - The response's http status code.

DeleteCorsPolicy

data DeleteCorsPolicy

See: newDeleteCorsPolicy smart constructor.

Constructors

DeleteCorsPolicy' Text

newDeleteCorsPolicy

Arguments

:: Text
$sel:containerName:DeleteCorsPolicy'

-> DeleteCorsPolicy

Create a value of DeleteCorsPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerName:DeleteCorsPolicy', deleteCorsPolicy_containerName - The name of the container to remove the policy from.

data DeleteCorsPolicyResponse

See: newDeleteCorsPolicyResponse smart constructor.

Constructors

DeleteCorsPolicyResponse' Int

newDeleteCorsPolicyResponse

Arguments

:: Int
$sel:httpStatus:DeleteCorsPolicyResponse'

-> DeleteCorsPolicyResponse

Create a value of DeleteCorsPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteCorsPolicyResponse', deleteCorsPolicyResponse_httpStatus - The response's http status code.

GetContainerPolicy

data GetContainerPolicy

See: newGetContainerPolicy smart constructor.

Constructors

GetContainerPolicy' Text

newGetContainerPolicy

Arguments

:: Text
$sel:containerName:GetContainerPolicy'

-> GetContainerPolicy

Create a value of GetContainerPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerName:GetContainerPolicy', getContainerPolicy_containerName - The name of the container.

data GetContainerPolicyResponse

See: newGetContainerPolicyResponse smart constructor.

Constructors

GetContainerPolicyResponse' Int Text

newGetContainerPolicyResponse

Arguments

:: Int
$sel:httpStatus:GetContainerPolicyResponse'

-> Text
$sel:policy:GetContainerPolicyResponse'

-> GetContainerPolicyResponse

Create a value of GetContainerPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetContainerPolicyResponse', getContainerPolicyResponse_httpStatus - The response's http status code.

$sel:policy:GetContainerPolicyResponse', getContainerPolicyResponse_policy - The contents of the access policy.

DeleteContainerPolicy

data DeleteContainerPolicy

See: newDeleteContainerPolicy smart constructor.

Constructors

DeleteContainerPolicy' Text

newDeleteContainerPolicy

Arguments

:: Text
$sel:containerName:DeleteContainerPolicy'

-> DeleteContainerPolicy

Create a value of DeleteContainerPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerName:DeleteContainerPolicy', deleteContainerPolicy_containerName - The name of the container that holds the policy.

data DeleteContainerPolicyResponse

See: newDeleteContainerPolicyResponse smart constructor.

Constructors

DeleteContainerPolicyResponse' Int

newDeleteContainerPolicyResponse

Arguments

:: Int
$sel:httpStatus:DeleteContainerPolicyResponse'

-> DeleteContainerPolicyResponse

Create a value of DeleteContainerPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteContainerPolicyResponse', deleteContainerPolicyResponse_httpStatus - The response's http status code.

ListContainers (Paginated)

data ListContainers

See: newListContainers smart constructor.

Constructors

ListContainers' (Maybe Text) (Maybe Natural)

newListContainers :: ListContainers

Create a value of ListContainers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListContainers', listContainers_nextToken - Only if you used MaxResults in the first command, enter the token (which was included in the previous response) to obtain the next set of containers. This token is included in a response only if there actually are more containers to list.

$sel:maxResults:ListContainers', listContainers_maxResults - Enter the maximum number of containers in the response. Use from 1 to 255 characters.

data ListContainersResponse

See: newListContainersResponse smart constructor.

Constructors

ListContainersResponse' (Maybe Text) Int [Container]

newListContainersResponse

Arguments

:: Int
$sel:httpStatus:ListContainersResponse'

-> ListContainersResponse

Create a value of ListContainersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListContainers', listContainersResponse_nextToken - NextToken is the token to use in the next call to ListContainers. This token is returned only if you included the MaxResults tag in the original command, and only if there are still containers to return.

$sel:httpStatus:ListContainersResponse', listContainersResponse_httpStatus - The response's http status code.

$sel:containers:ListContainersResponse', listContainersResponse_containers - The names of the containers.

CreateContainer

data CreateContainer

See: newCreateContainer smart constructor.

Constructors

CreateContainer' (Maybe (NonEmpty Tag)) Text

newCreateContainer

Arguments

:: Text
$sel:containerName:CreateContainer'

-> CreateContainer

Create a value of CreateContainer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateContainer', createContainer_tags - An array of key:value pairs that you define. These values can be anything that you want. Typically, the tag key represents a category (such as "environment") and the tag value represents a specific value within that category (such as "test," "development," or "production"). You can add up to 50 tags to each container. For more information about tagging, including naming and usage conventions, see Tagging Resources in MediaStore.

$sel:containerName:CreateContainer', createContainer_containerName - The name for the container. The name must be from 1 to 255 characters. Container names must be unique to your AWS account within a specific region. As an example, you could create a container named movies in every region, as long as you don’t have an existing container with that name.

data CreateContainerResponse

See: newCreateContainerResponse smart constructor.

Constructors

CreateContainerResponse' Int Container

newCreateContainerResponse

Arguments

:: Int
$sel:httpStatus:CreateContainerResponse'

-> Container
$sel:container:CreateContainerResponse'

-> CreateContainerResponse

Create a value of CreateContainerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateContainerResponse', createContainerResponse_httpStatus - The response's http status code.

$sel:container:CreateContainerResponse', createContainerResponse_container - ContainerARN: The Amazon Resource Name (ARN) of the newly created container. The ARN has the following format: arn:aws:<region>:<account that owns this container>:container/<name of container>. For example: arn:aws:mediastore:us-west-2:111122223333:container/movies

ContainerName: The container name as specified in the request.

CreationTime: Unix time stamp.

Status: The status of container creation or deletion. The status is one of the following: CREATING, ACTIVE, or DELETING. While the service is creating the container, the status is CREATING. When an endpoint is available, the status changes to ACTIVE.

The return value does not include the container's endpoint. To make downstream requests, you must obtain this value by using DescribeContainer or ListContainers.

GetLifecyclePolicy

data GetLifecyclePolicy

See: newGetLifecyclePolicy smart constructor.

Constructors

GetLifecyclePolicy' Text

newGetLifecyclePolicy

Arguments

:: Text
$sel:containerName:GetLifecyclePolicy'

-> GetLifecyclePolicy

Create a value of GetLifecyclePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerName:GetLifecyclePolicy', getLifecyclePolicy_containerName - The name of the container that the object lifecycle policy is assigned to.

data GetLifecyclePolicyResponse

See: newGetLifecyclePolicyResponse smart constructor.

Constructors

GetLifecyclePolicyResponse' Int Text

newGetLifecyclePolicyResponse

Arguments

:: Int
$sel:httpStatus:GetLifecyclePolicyResponse'

-> Text
$sel:lifecyclePolicy:GetLifecyclePolicyResponse'

-> GetLifecyclePolicyResponse

Create a value of GetLifecyclePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetLifecyclePolicyResponse', getLifecyclePolicyResponse_httpStatus - The response's http status code.

$sel:lifecyclePolicy:GetLifecyclePolicyResponse', getLifecyclePolicyResponse_lifecyclePolicy - The object lifecycle policy that is assigned to the container.

PutMetricPolicy

data PutMetricPolicy

See: newPutMetricPolicy smart constructor.

Constructors

PutMetricPolicy' Text MetricPolicy

newPutMetricPolicy

Arguments

:: Text
$sel:containerName:PutMetricPolicy'

-> MetricPolicy
$sel:metricPolicy:PutMetricPolicy'

-> PutMetricPolicy

Create a value of PutMetricPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:containerName:PutMetricPolicy', putMetricPolicy_containerName - The name of the container that you want to add the metric policy to.

$sel:metricPolicy:PutMetricPolicy', putMetricPolicy_metricPolicy - The metric policy that you want to associate with the container. In the policy, you must indicate whether you want MediaStore to send container-level metrics. You can also include up to five rules to define groups of objects that you want MediaStore to send object-level metrics for. If you include rules in the policy, construct each rule with both of the following:

An object group that defines which objects to include in the group. The definition can be a path or a file name, but it can't have more than 900 characters. Valid characters are: a-z, A-Z, 0-9, _ (underscore), = (equal), : (colon), . (period), - (hyphen), ~ (tilde), / (forward slash), and * (asterisk). Wildcards (*) are acceptable.

An object group name that allows you to refer to the object group. The name can't have more than 30 characters. Valid characters are: a-z, A-Z, 0-9, and _ (underscore).

data PutMetricPolicyResponse

See: newPutMetricPolicyResponse smart constructor.

Constructors

PutMetricPolicyResponse' Int

newPutMetricPolicyResponse

Arguments

:: Int
$sel:httpStatus:PutMetricPolicyResponse'

-> PutMetricPolicyResponse

Create a value of PutMetricPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutMetricPolicyResponse', putMetricPolicyResponse_httpStatus - The response's http status code.

ListTagsForResource

data ListTagsForResource

See: newListTagsForResource smart constructor.

Constructors

ListTagsForResource' Text

newListTagsForResource

Arguments

:: Text
$sel:resource:ListTagsForResource'

-> ListTagsForResource

Create a value of ListTagsForResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resource:ListTagsForResource', listTagsForResource_resource - The Amazon Resource Name (ARN) for the container.

data ListTagsForResourceResponse

See: newListTagsForResourceResponse smart constructor.

Constructors

ListTagsForResourceResponse' (Maybe (NonEmpty Tag)) Int

newListTagsForResourceResponse

Arguments

:: Int
$sel:httpStatus:ListTagsForResourceResponse'

-> ListTagsForResourceResponse

Create a value of ListTagsForResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:ListTagsForResourceResponse', listTagsForResourceResponse_tags - An array of key:value pairs that are assigned to the container.

$sel:httpStatus:ListTagsForResourceResponse', listTagsForResourceResponse_httpStatus - The response's http status code.

Types

ContainerLevelMetrics

newtype ContainerLevelMetrics

Constructors

ContainerLevelMetrics'

Fields

fromContainerLevelMetrics :: Text

Bundled Patterns

pattern ContainerLevelMetrics_DISABLED :: ContainerLevelMetrics

pattern ContainerLevelMetrics_ENABLED :: ContainerLevelMetrics

ContainerStatus

newtype ContainerStatus

Constructors

ContainerStatus'

Fields

fromContainerStatus :: Text

Bundled Patterns

pattern ContainerStatus_ACTIVE :: ContainerStatus

pattern ContainerStatus_CREATING :: ContainerStatus

pattern ContainerStatus_DELETING :: ContainerStatus

MethodName

newtype MethodName

Constructors

MethodName'

Fields

fromMethodName :: Text

Bundled Patterns

pattern MethodName_DELETE :: MethodName

pattern MethodName_GET :: MethodName

pattern MethodName_HEAD :: MethodName

pattern MethodName_PUT :: MethodName

Container

data Container

This section describes operations that you can perform on an AWS Elemental MediaStore container.

See: newContainer smart constructor.

Constructors

Container' (Maybe POSIX) (Maybe ContainerStatus) (Maybe Text) (Maybe Bool) (Maybe Text) (Maybe Text)

newContainer :: Container

Create a value of Container with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Container', container_creationTime - Unix timestamp.

$sel:status:Container', container_status - The status of container creation or deletion. The status is one of the following: CREATING, ACTIVE, or DELETING. While the service is creating the container, the status is CREATING. When the endpoint is available, the status changes to ACTIVE.

$sel:arn:Container', container_arn - The Amazon Resource Name (ARN) of the container. The ARN has the following format:

arn:aws:<region>:<account that owns this container>:container/<name of container>

For example: arn:aws:mediastore:us-west-2:111122223333:container/movies

$sel:accessLoggingEnabled:Container', container_accessLoggingEnabled - The state of access logging on the container. This value is false by default, indicating that AWS Elemental MediaStore does not send access logs to Amazon CloudWatch Logs. When you enable access logging on the container, MediaStore changes this value to true, indicating that the service delivers access logs for objects stored in that container to CloudWatch Logs.

$sel:name:Container', container_name - The name of the container.

$sel:endpoint:Container', container_endpoint - The DNS endpoint of the container. Use the endpoint to identify the specific container when sending requests to the data plane. The service assigns this value when the container is created. Once the value has been assigned, it does not change.

CorsRule

data CorsRule

A rule for a CORS policy. You can add up to 100 rules to a CORS policy. If more than one rule applies, the service uses the first applicable rule listed.

See: newCorsRule smart constructor.

Constructors

CorsRule' (Maybe (NonEmpty MethodName)) (Maybe Natural) (Maybe [Text]) (NonEmpty Text) [Text]

newCorsRule

Arguments

:: NonEmpty Text
$sel:allowedOrigins:CorsRule'

-> CorsRule

Create a value of CorsRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:allowedMethods:CorsRule', corsRule_allowedMethods - Identifies an HTTP method that the origin that is specified in the rule is allowed to execute.

Each CORS rule must contain at least one AllowedMethods and one AllowedOrigins element.

$sel:maxAgeSeconds:CorsRule', corsRule_maxAgeSeconds - The time in seconds that your browser caches the preflight response for the specified resource.

A CORS rule can have only one MaxAgeSeconds element.

$sel:exposeHeaders:CorsRule', corsRule_exposeHeaders - One or more headers in the response that you want users to be able to access from their applications (for example, from a JavaScript XMLHttpRequest object).

This element is optional for each rule.

$sel:allowedOrigins:CorsRule', corsRule_allowedOrigins - One or more response headers that you want users to be able to access from their applications (for example, from a JavaScript XMLHttpRequest object).

Each CORS rule must have at least one AllowedOrigins element. The string value can include only one wildcard character (*), for example, http://*.example.com. Additionally, you can specify only one wildcard character to allow cross-origin access for all origins.

$sel:allowedHeaders:CorsRule', corsRule_allowedHeaders - Specifies which headers are allowed in a preflight OPTIONS request through the Access-Control-Request-Headers header. Each header name that is specified in Access-Control-Request-Headers must have a corresponding entry in the rule. Only the headers that were requested are sent back.

This element can contain only one wildcard character (*).

MetricPolicy

data MetricPolicy

The metric policy that is associated with the container. A metric policy allows AWS Elemental MediaStore to send metrics to Amazon CloudWatch. In the policy, you must indicate whether you want MediaStore to send container-level metrics. You can also include rules to define groups of objects that you want MediaStore to send object-level metrics for.

To view examples of how to construct a metric policy for your use case, see Example Metric Policies.

See: newMetricPolicy smart constructor.

Constructors

MetricPolicy' (Maybe (NonEmpty MetricPolicyRule)) ContainerLevelMetrics

newMetricPolicy

Arguments

:: ContainerLevelMetrics
$sel:containerLevelMetrics:MetricPolicy'

-> MetricPolicy

Create a value of MetricPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricPolicyRules:MetricPolicy', metricPolicy_metricPolicyRules - A parameter that holds an array of rules that enable metrics at the object level. This parameter is optional, but if you choose to include it, you must also include at least one rule. By default, you can include up to five rules. You can also request a quota increase to allow up to 300 rules per policy.

$sel:containerLevelMetrics:MetricPolicy', metricPolicy_containerLevelMetrics - A setting to enable or disable metrics at the container level.

MetricPolicyRule

data MetricPolicyRule

A setting that enables metrics at the object level. Each rule contains an object group and an object group name. If the policy includes the MetricPolicyRules parameter, you must include at least one rule. Each metric policy can include up to five rules by default. You can also request a quota increase to allow up to 300 rules per policy.

See: newMetricPolicyRule smart constructor.

Constructors

MetricPolicyRule' Text Text

newMetricPolicyRule

Arguments

:: Text
$sel:objectGroup:MetricPolicyRule'

-> Text
$sel:objectGroupName:MetricPolicyRule'

-> MetricPolicyRule

Create a value of MetricPolicyRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:objectGroup:MetricPolicyRule', metricPolicyRule_objectGroup - A path or file name that defines which objects to include in the group. Wildcards (*) are acceptable.

$sel:objectGroupName:MetricPolicyRule', metricPolicyRule_objectGroupName - A name that allows you to refer to the object group.

Tag

data Tag

A collection of tags associated with a container. Each tag consists of a key:value pair, which can be anything you define. Typically, the tag key represents a category (such as "environment") and the tag value represents a specific value within that category (such as "test," "development," or "production"). You can add up to 50 tags to each container. For more information about tagging, including naming and usage conventions, see Tagging Resources in MediaStore.

See: newTag smart constructor.

Constructors

Tag' (Maybe Text) Text

newTag

Arguments

:: Text
$sel:key:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:Tag', tag_value - Part of the key:value pair that defines a tag. You can use a tag value to describe a specific value within a category, such as "companyA" or "companyB." Tag values are case-sensitive.

$sel:key:Tag', tag_key - Part of the key:value pair that defines a tag. You can use a tag key to describe a category of information, such as "customer." Tag keys are case-sensitive.