Network.AWS.MediaStore.Lens

Description

Operations

PutLifecyclePolicy

putLifecyclePolicy_containerName :: Lens' PutLifecyclePolicy Text

The name of the container that you want to assign the object lifecycle policy to.

putLifecyclePolicy_lifecyclePolicy :: Lens' PutLifecyclePolicy Text

The object lifecycle policy to apply to the container.

putLifecyclePolicyResponse_httpStatus :: Lens' PutLifecyclePolicyResponse Int

The response's http status code.

PutCorsPolicy

putCorsPolicy_containerName :: Lens' PutCorsPolicy Text

The name of the container that you want to assign the CORS policy to.

putCorsPolicy_corsPolicy :: Lens' PutCorsPolicy (NonEmpty CorsRule)

The CORS policy to apply to the container.

putCorsPolicyResponse_httpStatus :: Lens' PutCorsPolicyResponse Int

The response's http status code.

DeleteContainer

deleteContainer_containerName :: Lens' DeleteContainer Text

The name of the container to delete.

deleteContainerResponse_httpStatus :: Lens' DeleteContainerResponse Int

The response's http status code.

PutContainerPolicy

putContainerPolicy_containerName :: Lens' PutContainerPolicy Text

The name of the container.

putContainerPolicy_policy :: Lens' PutContainerPolicy Text

The contents of the policy, which includes the following:

One Version tag

One Statement tag that contains the standard tags for the policy.

putContainerPolicyResponse_httpStatus :: Lens' PutContainerPolicyResponse Int

The response's http status code.

GetCorsPolicy

getCorsPolicy_containerName :: Lens' GetCorsPolicy Text

The name of the container that the policy is assigned to.

getCorsPolicyResponse_httpStatus :: Lens' GetCorsPolicyResponse Int

The response's http status code.

getCorsPolicyResponse_corsPolicy :: Lens' GetCorsPolicyResponse (NonEmpty CorsRule)

The CORS policy assigned to the container.

UntagResource

untagResource_resource :: Lens' UntagResource Text

The Amazon Resource Name (ARN) for the container.

untagResource_tagKeys :: Lens' UntagResource [Text]

A comma-separated list of keys for tags that you want to remove from the container. For example, if your container has two tags (customer:CompanyA and priority:High) and you want to remove one of the tags (priority:High), you specify the key for the tag that you want to remove (priority).

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

TagResource

tagResource_resource :: Lens' TagResource Text

The Amazon Resource Name (ARN) for the container.

tagResource_tags :: Lens' TagResource (NonEmpty Tag)

An array of key:value pairs that you want to add to the container. You need to specify only the tags that you want to add or update. For example, suppose a container already has two tags (customer:CompanyA and priority:High). You want to change the priority tag and also add a third tag (type:Contract). For TagResource, you specify the following tags: priority:Medium, type:Contract. The result is that your container has three tags: customer:CompanyA, priority:Medium, and type:Contract.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

DeleteMetricPolicy

deleteMetricPolicy_containerName :: Lens' DeleteMetricPolicy Text

The name of the container that is associated with the metric policy that you want to delete.

deleteMetricPolicyResponse_httpStatus :: Lens' DeleteMetricPolicyResponse Int

The response's http status code.

DescribeContainer

describeContainer_containerName :: Lens' DescribeContainer (Maybe Text)

The name of the container to query.

describeContainerResponse_container :: Lens' DescribeContainerResponse (Maybe Container)

The name of the queried container.

describeContainerResponse_httpStatus :: Lens' DescribeContainerResponse Int

The response's http status code.

GetMetricPolicy

getMetricPolicy_containerName :: Lens' GetMetricPolicy Text

The name of the container that is associated with the metric policy.

getMetricPolicyResponse_httpStatus :: Lens' GetMetricPolicyResponse Int

The response's http status code.

getMetricPolicyResponse_metricPolicy :: Lens' GetMetricPolicyResponse MetricPolicy

The metric policy that is associated with the specific container.

StopAccessLogging

stopAccessLogging_containerName :: Lens' StopAccessLogging Text

The name of the container that you want to stop access logging on.

stopAccessLoggingResponse_httpStatus :: Lens' StopAccessLoggingResponse Int

The response's http status code.

StartAccessLogging

startAccessLogging_containerName :: Lens' StartAccessLogging Text

The name of the container that you want to start access logging on.

startAccessLoggingResponse_httpStatus :: Lens' StartAccessLoggingResponse Int

The response's http status code.

DeleteLifecyclePolicy

deleteLifecyclePolicy_containerName :: Lens' DeleteLifecyclePolicy Text

The name of the container that holds the object lifecycle policy.

deleteLifecyclePolicyResponse_httpStatus :: Lens' DeleteLifecyclePolicyResponse Int

The response's http status code.

DeleteCorsPolicy

deleteCorsPolicy_containerName :: Lens' DeleteCorsPolicy Text

The name of the container to remove the policy from.

deleteCorsPolicyResponse_httpStatus :: Lens' DeleteCorsPolicyResponse Int

The response's http status code.

GetContainerPolicy

getContainerPolicy_containerName :: Lens' GetContainerPolicy Text

The name of the container.

getContainerPolicyResponse_httpStatus :: Lens' GetContainerPolicyResponse Int

The response's http status code.

getContainerPolicyResponse_policy :: Lens' GetContainerPolicyResponse Text

The contents of the access policy.

DeleteContainerPolicy

deleteContainerPolicy_containerName :: Lens' DeleteContainerPolicy Text

The name of the container that holds the policy.

deleteContainerPolicyResponse_httpStatus :: Lens' DeleteContainerPolicyResponse Int

The response's http status code.

ListContainers

listContainers_nextToken :: Lens' ListContainers (Maybe Text)

Only if you used MaxResults in the first command, enter the token (which was included in the previous response) to obtain the next set of containers. This token is included in a response only if there actually are more containers to list.

listContainers_maxResults :: Lens' ListContainers (Maybe Natural)

Enter the maximum number of containers in the response. Use from 1 to 255 characters.

listContainersResponse_nextToken :: Lens' ListContainersResponse (Maybe Text)

NextToken is the token to use in the next call to ListContainers. This token is returned only if you included the MaxResults tag in the original command, and only if there are still containers to return.

listContainersResponse_httpStatus :: Lens' ListContainersResponse Int

The response's http status code.

listContainersResponse_containers :: Lens' ListContainersResponse [Container]

The names of the containers.

CreateContainer

createContainer_tags :: Lens' CreateContainer (Maybe (NonEmpty Tag))

An array of key:value pairs that you define. These values can be anything that you want. Typically, the tag key represents a category (such as "environment") and the tag value represents a specific value within that category (such as "test," "development," or "production"). You can add up to 50 tags to each container. For more information about tagging, including naming and usage conventions, see Tagging Resources in MediaStore.

createContainer_containerName :: Lens' CreateContainer Text

The name for the container. The name must be from 1 to 255 characters. Container names must be unique to your AWS account within a specific region. As an example, you could create a container named movies in every region, as long as you don’t have an existing container with that name.

createContainerResponse_httpStatus :: Lens' CreateContainerResponse Int

The response's http status code.

createContainerResponse_container :: Lens' CreateContainerResponse Container

ContainerARN: The Amazon Resource Name (ARN) of the newly created container. The ARN has the following format: arn:aws:<region>:<account that owns this container>:container/<name of container>. For example: arn:aws:mediastore:us-west-2:111122223333:container/movies

ContainerName: The container name as specified in the request.

CreationTime: Unix time stamp.

Status: The status of container creation or deletion. The status is one of the following: CREATING, ACTIVE, or DELETING. While the service is creating the container, the status is CREATING. When an endpoint is available, the status changes to ACTIVE.

The return value does not include the container's endpoint. To make downstream requests, you must obtain this value by using DescribeContainer or ListContainers.

GetLifecyclePolicy

getLifecyclePolicy_containerName :: Lens' GetLifecyclePolicy Text

The name of the container that the object lifecycle policy is assigned to.

getLifecyclePolicyResponse_httpStatus :: Lens' GetLifecyclePolicyResponse Int

The response's http status code.

getLifecyclePolicyResponse_lifecyclePolicy :: Lens' GetLifecyclePolicyResponse Text

The object lifecycle policy that is assigned to the container.

PutMetricPolicy

putMetricPolicy_containerName :: Lens' PutMetricPolicy Text

The name of the container that you want to add the metric policy to.

putMetricPolicy_metricPolicy :: Lens' PutMetricPolicy MetricPolicy

The metric policy that you want to associate with the container. In the policy, you must indicate whether you want MediaStore to send container-level metrics. You can also include up to five rules to define groups of objects that you want MediaStore to send object-level metrics for. If you include rules in the policy, construct each rule with both of the following:

An object group that defines which objects to include in the group. The definition can be a path or a file name, but it can't have more than 900 characters. Valid characters are: a-z, A-Z, 0-9, _ (underscore), = (equal), : (colon), . (period), - (hyphen), ~ (tilde), / (forward slash), and * (asterisk). Wildcards (*) are acceptable.

An object group name that allows you to refer to the object group. The name can't have more than 30 characters. Valid characters are: a-z, A-Z, 0-9, and _ (underscore).

putMetricPolicyResponse_httpStatus :: Lens' PutMetricPolicyResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resource :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) for the container.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (NonEmpty Tag))

An array of key:value pairs that are assigned to the container.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

Types

Container

container_creationTime :: Lens' Container (Maybe UTCTime)

Unix timestamp.

container_status :: Lens' Container (Maybe ContainerStatus)

The status of container creation or deletion. The status is one of the following: CREATING, ACTIVE, or DELETING. While the service is creating the container, the status is CREATING. When the endpoint is available, the status changes to ACTIVE.

container_arn :: Lens' Container (Maybe Text)

The Amazon Resource Name (ARN) of the container. The ARN has the following format:

arn:aws:<region>:<account that owns this container>:container/<name of container>

For example: arn:aws:mediastore:us-west-2:111122223333:container/movies

container_accessLoggingEnabled :: Lens' Container (Maybe Bool)

The state of access logging on the container. This value is false by default, indicating that AWS Elemental MediaStore does not send access logs to Amazon CloudWatch Logs. When you enable access logging on the container, MediaStore changes this value to true, indicating that the service delivers access logs for objects stored in that container to CloudWatch Logs.

container_name :: Lens' Container (Maybe Text)

The name of the container.

container_endpoint :: Lens' Container (Maybe Text)

The DNS endpoint of the container. Use the endpoint to identify the specific container when sending requests to the data plane. The service assigns this value when the container is created. Once the value has been assigned, it does not change.

CorsRule

corsRule_allowedMethods :: Lens' CorsRule (Maybe (NonEmpty MethodName))

Identifies an HTTP method that the origin that is specified in the rule is allowed to execute.

Each CORS rule must contain at least one AllowedMethods and one AllowedOrigins element.

corsRule_maxAgeSeconds :: Lens' CorsRule (Maybe Natural)

The time in seconds that your browser caches the preflight response for the specified resource.

A CORS rule can have only one MaxAgeSeconds element.

corsRule_exposeHeaders :: Lens' CorsRule (Maybe [Text])

One or more headers in the response that you want users to be able to access from their applications (for example, from a JavaScript XMLHttpRequest object).

This element is optional for each rule.

corsRule_allowedOrigins :: Lens' CorsRule (NonEmpty Text)

One or more response headers that you want users to be able to access from their applications (for example, from a JavaScript XMLHttpRequest object).

Each CORS rule must have at least one AllowedOrigins element. The string value can include only one wildcard character (*), for example, http://*.example.com. Additionally, you can specify only one wildcard character to allow cross-origin access for all origins.

corsRule_allowedHeaders :: Lens' CorsRule [Text]

Specifies which headers are allowed in a preflight OPTIONS request through the Access-Control-Request-Headers header. Each header name that is specified in Access-Control-Request-Headers must have a corresponding entry in the rule. Only the headers that were requested are sent back.

This element can contain only one wildcard character (*).

MetricPolicy

metricPolicy_metricPolicyRules :: Lens' MetricPolicy (Maybe (NonEmpty MetricPolicyRule))

A parameter that holds an array of rules that enable metrics at the object level. This parameter is optional, but if you choose to include it, you must also include at least one rule. By default, you can include up to five rules. You can also request a quota increase to allow up to 300 rules per policy.

metricPolicy_containerLevelMetrics :: Lens' MetricPolicy ContainerLevelMetrics

A setting to enable or disable metrics at the container level.

MetricPolicyRule

metricPolicyRule_objectGroup :: Lens' MetricPolicyRule Text

A path or file name that defines which objects to include in the group. Wildcards (*) are acceptable.

metricPolicyRule_objectGroupName :: Lens' MetricPolicyRule Text

A name that allows you to refer to the object group.

Tag

tag_value :: Lens' Tag (Maybe Text)

Part of the key:value pair that defines a tag. You can use a tag value to describe a specific value within a category, such as "companyA" or "companyB." Tag values are case-sensitive.

tag_key :: Lens' Tag Text

Part of the key:value pair that defines a tag. You can use a tag key to describe a category of information, such as "customer." Tag keys are case-sensitive.