Network.AWS.MediaStore.Types

Description

Service Configuration

defaultService :: Service

API version 2017-09-01 of the Amazon Elemental MediaStore SDK configuration.

Errors

_ContainerInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The container that you specified in the request already exists or is being updated.

_InternalServerError :: AsError a => Getting (First ServiceError) a ServiceError

The service is temporarily unavailable.

_PolicyNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The policy that you specified in the request does not exist.

_CorsPolicyNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The CORS policy that you specified in the request does not exist.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

A service limit has been exceeded.

_ContainerNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The container that you specified in the request does not exist.

ContainerLevelMetrics

newtype ContainerLevelMetrics

Constructors

ContainerLevelMetrics'

Fields

fromContainerLevelMetrics :: Text

Bundled Patterns

pattern ContainerLevelMetrics_DISABLED :: ContainerLevelMetrics

pattern ContainerLevelMetrics_ENABLED :: ContainerLevelMetrics

ContainerStatus

newtype ContainerStatus

Constructors

ContainerStatus'

Fields

fromContainerStatus :: Text

Bundled Patterns

pattern ContainerStatus_ACTIVE :: ContainerStatus

pattern ContainerStatus_CREATING :: ContainerStatus

pattern ContainerStatus_DELETING :: ContainerStatus

MethodName

newtype MethodName

Constructors

MethodName'

Fields

fromMethodName :: Text

Bundled Patterns

pattern MethodName_DELETE :: MethodName

pattern MethodName_GET :: MethodName

pattern MethodName_HEAD :: MethodName

pattern MethodName_PUT :: MethodName

Container

data Container

This section describes operations that you can perform on an AWS Elemental MediaStore container.

See: newContainer smart constructor.

Constructors

Container'

Fields

creationTime :: Maybe POSIX

Unix timestamp.

status :: Maybe ContainerStatus

The status of container creation or deletion. The status is one of the following: CREATING, ACTIVE, or DELETING. While the service is creating the container, the status is CREATING. When the endpoint is available, the status changes to ACTIVE.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the container. The ARN has the following format:

arn:aws:<region>:<account that owns this container>:container/<name of container>

For example: arn:aws:mediastore:us-west-2:111122223333:container/movies

accessLoggingEnabled :: Maybe Bool

The state of access logging on the container. This value is false by default, indicating that AWS Elemental MediaStore does not send access logs to Amazon CloudWatch Logs. When you enable access logging on the container, MediaStore changes this value to true, indicating that the service delivers access logs for objects stored in that container to CloudWatch Logs.

name :: Maybe Text

The name of the container.

endpoint :: Maybe Text

The DNS endpoint of the container. Use the endpoint to identify the specific container when sending requests to the data plane. The service assigns this value when the container is created. Once the value has been assigned, it does not change.

newContainer :: Container

Create a value of Container with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Container', container_creationTime - Unix timestamp.

$sel:status:Container', container_status - The status of container creation or deletion. The status is one of the following: CREATING, ACTIVE, or DELETING. While the service is creating the container, the status is CREATING. When the endpoint is available, the status changes to ACTIVE.

$sel:arn:Container', container_arn - The Amazon Resource Name (ARN) of the container. The ARN has the following format:

arn:aws:<region>:<account that owns this container>:container/<name of container>

For example: arn:aws:mediastore:us-west-2:111122223333:container/movies

$sel:accessLoggingEnabled:Container', container_accessLoggingEnabled - The state of access logging on the container. This value is false by default, indicating that AWS Elemental MediaStore does not send access logs to Amazon CloudWatch Logs. When you enable access logging on the container, MediaStore changes this value to true, indicating that the service delivers access logs for objects stored in that container to CloudWatch Logs.

$sel:name:Container', container_name - The name of the container.

$sel:endpoint:Container', container_endpoint - The DNS endpoint of the container. Use the endpoint to identify the specific container when sending requests to the data plane. The service assigns this value when the container is created. Once the value has been assigned, it does not change.

container_creationTime :: Lens' Container (Maybe UTCTime)

Unix timestamp.

container_status :: Lens' Container (Maybe ContainerStatus)

The status of container creation or deletion. The status is one of the following: CREATING, ACTIVE, or DELETING. While the service is creating the container, the status is CREATING. When the endpoint is available, the status changes to ACTIVE.

container_arn :: Lens' Container (Maybe Text)

The Amazon Resource Name (ARN) of the container. The ARN has the following format:

arn:aws:<region>:<account that owns this container>:container/<name of container>

For example: arn:aws:mediastore:us-west-2:111122223333:container/movies

container_accessLoggingEnabled :: Lens' Container (Maybe Bool)

The state of access logging on the container. This value is false by default, indicating that AWS Elemental MediaStore does not send access logs to Amazon CloudWatch Logs. When you enable access logging on the container, MediaStore changes this value to true, indicating that the service delivers access logs for objects stored in that container to CloudWatch Logs.

container_name :: Lens' Container (Maybe Text)

The name of the container.

container_endpoint :: Lens' Container (Maybe Text)

The DNS endpoint of the container. Use the endpoint to identify the specific container when sending requests to the data plane. The service assigns this value when the container is created. Once the value has been assigned, it does not change.

CorsRule

data CorsRule

A rule for a CORS policy. You can add up to 100 rules to a CORS policy. If more than one rule applies, the service uses the first applicable rule listed.

See: newCorsRule smart constructor.

Constructors

CorsRule'

Fields

allowedMethods :: Maybe (NonEmpty MethodName)

Identifies an HTTP method that the origin that is specified in the rule is allowed to execute.

Each CORS rule must contain at least one AllowedMethods and one AllowedOrigins element.

maxAgeSeconds :: Maybe Natural

The time in seconds that your browser caches the preflight response for the specified resource.

A CORS rule can have only one MaxAgeSeconds element.

exposeHeaders :: Maybe [Text]

One or more headers in the response that you want users to be able to access from their applications (for example, from a JavaScript XMLHttpRequest object).

This element is optional for each rule.

allowedOrigins :: NonEmpty Text

One or more response headers that you want users to be able to access from their applications (for example, from a JavaScript XMLHttpRequest object).

Each CORS rule must have at least one AllowedOrigins element. The string value can include only one wildcard character (*), for example, http://*.example.com. Additionally, you can specify only one wildcard character to allow cross-origin access for all origins.

allowedHeaders :: [Text]

Specifies which headers are allowed in a preflight OPTIONS request through the Access-Control-Request-Headers header. Each header name that is specified in Access-Control-Request-Headers must have a corresponding entry in the rule. Only the headers that were requested are sent back.

This element can contain only one wildcard character (*).

newCorsRule

Arguments

:: NonEmpty Text
$sel:allowedOrigins:CorsRule'

-> CorsRule

Create a value of CorsRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:allowedMethods:CorsRule', corsRule_allowedMethods - Identifies an HTTP method that the origin that is specified in the rule is allowed to execute.

Each CORS rule must contain at least one AllowedMethods and one AllowedOrigins element.

$sel:maxAgeSeconds:CorsRule', corsRule_maxAgeSeconds - The time in seconds that your browser caches the preflight response for the specified resource.

A CORS rule can have only one MaxAgeSeconds element.

$sel:exposeHeaders:CorsRule', corsRule_exposeHeaders - One or more headers in the response that you want users to be able to access from their applications (for example, from a JavaScript XMLHttpRequest object).

This element is optional for each rule.

$sel:allowedOrigins:CorsRule', corsRule_allowedOrigins - One or more response headers that you want users to be able to access from their applications (for example, from a JavaScript XMLHttpRequest object).

Each CORS rule must have at least one AllowedOrigins element. The string value can include only one wildcard character (*), for example, http://*.example.com. Additionally, you can specify only one wildcard character to allow cross-origin access for all origins.

$sel:allowedHeaders:CorsRule', corsRule_allowedHeaders - Specifies which headers are allowed in a preflight OPTIONS request through the Access-Control-Request-Headers header. Each header name that is specified in Access-Control-Request-Headers must have a corresponding entry in the rule. Only the headers that were requested are sent back.

This element can contain only one wildcard character (*).

corsRule_allowedMethods :: Lens' CorsRule (Maybe (NonEmpty MethodName))

Identifies an HTTP method that the origin that is specified in the rule is allowed to execute.

Each CORS rule must contain at least one AllowedMethods and one AllowedOrigins element.

corsRule_maxAgeSeconds :: Lens' CorsRule (Maybe Natural)

The time in seconds that your browser caches the preflight response for the specified resource.

A CORS rule can have only one MaxAgeSeconds element.

corsRule_exposeHeaders :: Lens' CorsRule (Maybe [Text])

One or more headers in the response that you want users to be able to access from their applications (for example, from a JavaScript XMLHttpRequest object).

This element is optional for each rule.

corsRule_allowedOrigins :: Lens' CorsRule (NonEmpty Text)

One or more response headers that you want users to be able to access from their applications (for example, from a JavaScript XMLHttpRequest object).

Each CORS rule must have at least one AllowedOrigins element. The string value can include only one wildcard character (*), for example, http://*.example.com. Additionally, you can specify only one wildcard character to allow cross-origin access for all origins.

corsRule_allowedHeaders :: Lens' CorsRule [Text]

Specifies which headers are allowed in a preflight OPTIONS request through the Access-Control-Request-Headers header. Each header name that is specified in Access-Control-Request-Headers must have a corresponding entry in the rule. Only the headers that were requested are sent back.

This element can contain only one wildcard character (*).

MetricPolicy

data MetricPolicy

The metric policy that is associated with the container. A metric policy allows AWS Elemental MediaStore to send metrics to Amazon CloudWatch. In the policy, you must indicate whether you want MediaStore to send container-level metrics. You can also include rules to define groups of objects that you want MediaStore to send object-level metrics for.

To view examples of how to construct a metric policy for your use case, see Example Metric Policies.

See: newMetricPolicy smart constructor.

Constructors

MetricPolicy'

Fields

metricPolicyRules :: Maybe (NonEmpty MetricPolicyRule)

A parameter that holds an array of rules that enable metrics at the object level. This parameter is optional, but if you choose to include it, you must also include at least one rule. By default, you can include up to five rules. You can also request a quota increase to allow up to 300 rules per policy.

containerLevelMetrics :: ContainerLevelMetrics

A setting to enable or disable metrics at the container level.

newMetricPolicy

Arguments

:: ContainerLevelMetrics
$sel:containerLevelMetrics:MetricPolicy'

-> MetricPolicy

Create a value of MetricPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricPolicyRules:MetricPolicy', metricPolicy_metricPolicyRules - A parameter that holds an array of rules that enable metrics at the object level. This parameter is optional, but if you choose to include it, you must also include at least one rule. By default, you can include up to five rules. You can also request a quota increase to allow up to 300 rules per policy.

$sel:containerLevelMetrics:MetricPolicy', metricPolicy_containerLevelMetrics - A setting to enable or disable metrics at the container level.

metricPolicy_metricPolicyRules :: Lens' MetricPolicy (Maybe (NonEmpty MetricPolicyRule))

A parameter that holds an array of rules that enable metrics at the object level. This parameter is optional, but if you choose to include it, you must also include at least one rule. By default, you can include up to five rules. You can also request a quota increase to allow up to 300 rules per policy.

metricPolicy_containerLevelMetrics :: Lens' MetricPolicy ContainerLevelMetrics

A setting to enable or disable metrics at the container level.

MetricPolicyRule

data MetricPolicyRule

A setting that enables metrics at the object level. Each rule contains an object group and an object group name. If the policy includes the MetricPolicyRules parameter, you must include at least one rule. Each metric policy can include up to five rules by default. You can also request a quota increase to allow up to 300 rules per policy.

See: newMetricPolicyRule smart constructor.

Constructors

MetricPolicyRule'

Fields

objectGroup :: Text

A path or file name that defines which objects to include in the group. Wildcards (*) are acceptable.

objectGroupName :: Text

A name that allows you to refer to the object group.

newMetricPolicyRule

Arguments

:: Text
$sel:objectGroup:MetricPolicyRule'

-> Text
$sel:objectGroupName:MetricPolicyRule'

-> MetricPolicyRule

Create a value of MetricPolicyRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:objectGroup:MetricPolicyRule', metricPolicyRule_objectGroup - A path or file name that defines which objects to include in the group. Wildcards (*) are acceptable.

$sel:objectGroupName:MetricPolicyRule', metricPolicyRule_objectGroupName - A name that allows you to refer to the object group.

metricPolicyRule_objectGroup :: Lens' MetricPolicyRule Text

A path or file name that defines which objects to include in the group. Wildcards (*) are acceptable.

metricPolicyRule_objectGroupName :: Lens' MetricPolicyRule Text

A name that allows you to refer to the object group.

Tag

data Tag

A collection of tags associated with a container. Each tag consists of a key:value pair, which can be anything you define. Typically, the tag key represents a category (such as "environment") and the tag value represents a specific value within that category (such as "test," "development," or "production"). You can add up to 50 tags to each container. For more information about tagging, including naming and usage conventions, see Tagging Resources in MediaStore.

See: newTag smart constructor.

Constructors

Tag'

Fields

value :: Maybe Text

Part of the key:value pair that defines a tag. You can use a tag value to describe a specific value within a category, such as "companyA" or "companyB." Tag values are case-sensitive.

key :: Text

Part of the key:value pair that defines a tag. You can use a tag key to describe a category of information, such as "customer." Tag keys are case-sensitive.

newTag

Arguments

:: Text
$sel:key:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:Tag', tag_value - Part of the key:value pair that defines a tag. You can use a tag value to describe a specific value within a category, such as "companyA" or "companyB." Tag values are case-sensitive.

$sel:key:Tag', tag_key - Part of the key:value pair that defines a tag. You can use a tag key to describe a category of information, such as "customer." Tag keys are case-sensitive.

tag_value :: Lens' Tag (Maybe Text)

Part of the key:value pair that defines a tag. You can use a tag value to describe a specific value within a category, such as "companyA" or "companyB." Tag values are case-sensitive.

tag_key :: Lens' Tag Text

Part of the key:value pair that defines a tag. You can use a tag key to describe a category of information, such as "customer." Tag keys are case-sensitive.