Network.AWS.MediaStore.Types.MetricPolicy

Description

Documentation

data MetricPolicy

The metric policy that is associated with the container. A metric policy allows AWS Elemental MediaStore to send metrics to Amazon CloudWatch. In the policy, you must indicate whether you want MediaStore to send container-level metrics. You can also include rules to define groups of objects that you want MediaStore to send object-level metrics for.

To view examples of how to construct a metric policy for your use case, see Example Metric Policies.

See: newMetricPolicy smart constructor.

Constructors

MetricPolicy'

Fields

metricPolicyRules :: Maybe (NonEmpty MetricPolicyRule)

A parameter that holds an array of rules that enable metrics at the object level. This parameter is optional, but if you choose to include it, you must also include at least one rule. By default, you can include up to five rules. You can also request a quota increase to allow up to 300 rules per policy.

containerLevelMetrics :: ContainerLevelMetrics

A setting to enable or disable metrics at the container level.

newMetricPolicy

Arguments

:: ContainerLevelMetrics
$sel:containerLevelMetrics:MetricPolicy'

-> MetricPolicy

Create a value of MetricPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metricPolicyRules:MetricPolicy', metricPolicy_metricPolicyRules - A parameter that holds an array of rules that enable metrics at the object level. This parameter is optional, but if you choose to include it, you must also include at least one rule. By default, you can include up to five rules. You can also request a quota increase to allow up to 300 rules per policy.

$sel:containerLevelMetrics:MetricPolicy', metricPolicy_containerLevelMetrics - A setting to enable or disable metrics at the container level.

metricPolicy_metricPolicyRules :: Lens' MetricPolicy (Maybe (NonEmpty MetricPolicyRule))

A parameter that holds an array of rules that enable metrics at the object level. This parameter is optional, but if you choose to include it, you must also include at least one rule. By default, you can include up to five rules. You can also request a quota increase to allow up to 300 rules per policy.

metricPolicy_containerLevelMetrics :: Lens' MetricPolicy ContainerLevelMetrics

A setting to enable or disable metrics at the container level.