Network.AWS.MediaStore.Types.MetricPolicyRule

Description

Documentation

data MetricPolicyRule

A setting that enables metrics at the object level. Each rule contains an object group and an object group name. If the policy includes the MetricPolicyRules parameter, you must include at least one rule. Each metric policy can include up to five rules by default. You can also request a quota increase to allow up to 300 rules per policy.

See: newMetricPolicyRule smart constructor.

Constructors

MetricPolicyRule'

Fields

objectGroup :: Text

A path or file name that defines which objects to include in the group. Wildcards (*) are acceptable.

objectGroupName :: Text

A name that allows you to refer to the object group.

newMetricPolicyRule

Arguments

:: Text
$sel:objectGroup:MetricPolicyRule'

-> Text
$sel:objectGroupName:MetricPolicyRule'

-> MetricPolicyRule

Create a value of MetricPolicyRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:objectGroup:MetricPolicyRule', metricPolicyRule_objectGroup - A path or file name that defines which objects to include in the group. Wildcards (*) are acceptable.

$sel:objectGroupName:MetricPolicyRule', metricPolicyRule_objectGroupName - A name that allows you to refer to the object group.

metricPolicyRule_objectGroup :: Lens' MetricPolicyRule Text

A path or file name that defines which objects to include in the group. Wildcards (*) are acceptable.

metricPolicyRule_objectGroupName :: Lens' MetricPolicyRule Text

A name that allows you to refer to the object group.