Network.AWS.MediaStoreData.DescribeObject

Description

Gets the headers for an object at the specified path.

Creating a Request

data DescribeObject

See: newDescribeObject smart constructor.

Constructors

DescribeObject'

Fields

path :: Text

The path (including the file name) where the object is stored in the container. Format: <folder name>/<folder name>/<file name>

newDescribeObject

Arguments

:: Text
$sel:path:DescribeObject'

-> DescribeObject

Create a value of DescribeObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:path:DescribeObject', describeObject_path - The path (including the file name) where the object is stored in the container. Format: <folder name>/<folder name>/<file name>

Request Lenses

describeObject_path :: Lens' DescribeObject Text

The path (including the file name) where the object is stored in the container. Format: <folder name>/<folder name>/<file name>

Destructuring the Response

data DescribeObjectResponse

See: newDescribeObjectResponse smart constructor.

Constructors

DescribeObjectResponse'

Fields

eTag :: Maybe Text

The ETag that represents a unique instance of the object.

contentLength :: Maybe Natural

The length of the object in bytes.

cacheControl :: Maybe Text

An optional CacheControl header that allows the caller to control the object's cache behavior. Headers can be passed in as specified in the HTTP at https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.9.

Headers with a custom user-defined value are also accepted.

lastModified :: Maybe POSIX

The date and time that the object was last modified.

contentType :: Maybe Text

The content type of the object.

httpStatus :: Int

The response's http status code.

newDescribeObjectResponse

Arguments

:: Int
$sel:httpStatus:DescribeObjectResponse'

-> DescribeObjectResponse

Create a value of DescribeObjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:DescribeObjectResponse', describeObjectResponse_eTag - The ETag that represents a unique instance of the object.

$sel:contentLength:DescribeObjectResponse', describeObjectResponse_contentLength - The length of the object in bytes.

$sel:cacheControl:DescribeObjectResponse', describeObjectResponse_cacheControl - An optional CacheControl header that allows the caller to control the object's cache behavior. Headers can be passed in as specified in the HTTP at https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.9.

Headers with a custom user-defined value are also accepted.

$sel:lastModified:DescribeObjectResponse', describeObjectResponse_lastModified - The date and time that the object was last modified.

$sel:contentType:DescribeObjectResponse', describeObjectResponse_contentType - The content type of the object.

$sel:httpStatus:DescribeObjectResponse', describeObjectResponse_httpStatus - The response's http status code.

Response Lenses

describeObjectResponse_eTag :: Lens' DescribeObjectResponse (Maybe Text)

The ETag that represents a unique instance of the object.

describeObjectResponse_contentLength :: Lens' DescribeObjectResponse (Maybe Natural)

The length of the object in bytes.

describeObjectResponse_cacheControl :: Lens' DescribeObjectResponse (Maybe Text)

An optional CacheControl header that allows the caller to control the object's cache behavior. Headers can be passed in as specified in the HTTP at https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.9.

Headers with a custom user-defined value are also accepted.

describeObjectResponse_lastModified :: Lens' DescribeObjectResponse (Maybe UTCTime)

The date and time that the object was last modified.

describeObjectResponse_contentType :: Lens' DescribeObjectResponse (Maybe Text)

The content type of the object.

describeObjectResponse_httpStatus :: Lens' DescribeObjectResponse Int

The response's http status code.