Network.AWS.MediaStoreData.Lens

Description

Operations

ListItems

listItems_nextToken :: Lens' ListItems (Maybe Text)

The token that identifies which batch of results that you want to see. For example, you submit a ListItems request with MaxResults set at 500. The service returns the first batch of results (up to 500) and a NextToken value. To see the next batch of results, you can submit the ListItems request a second time and specify the NextToken value.

Tokens expire after 15 minutes.

listItems_maxResults :: Lens' ListItems (Maybe Natural)

The maximum number of results to return per API request. For example, you submit a ListItems request with MaxResults set at 500. Although 2,000 items match your request, the service returns no more than the first 500 items. (The service also returns a NextToken value that you can use to fetch the next batch of results.) The service might return fewer results than the MaxResults value.

If MaxResults is not included in the request, the service defaults to pagination with a maximum of 1,000 results per page.

listItems_path :: Lens' ListItems (Maybe Text)

The path in the container from which to retrieve items. Format: <folder name>/<folder name>/<file name>

listItemsResponse_nextToken :: Lens' ListItemsResponse (Maybe Text)

The token that can be used in a request to view the next set of results. For example, you submit a ListItems request that matches 2,000 items with MaxResults set at 500. The service returns the first batch of results (up to 500) and a NextToken value that can be used to fetch the next batch of results.

listItemsResponse_items :: Lens' ListItemsResponse (Maybe [Item])

The metadata entries for the folders and objects at the requested path.

listItemsResponse_httpStatus :: Lens' ListItemsResponse Int

The response's http status code.

DeleteObject

deleteObject_path :: Lens' DeleteObject Text

The path (including the file name) where the object is stored in the container. Format: <folder name>/<folder name>/<file name>

deleteObjectResponse_httpStatus :: Lens' DeleteObjectResponse Int

The response's http status code.

DescribeObject

describeObject_path :: Lens' DescribeObject Text

The path (including the file name) where the object is stored in the container. Format: <folder name>/<folder name>/<file name>

describeObjectResponse_eTag :: Lens' DescribeObjectResponse (Maybe Text)

The ETag that represents a unique instance of the object.

describeObjectResponse_contentType :: Lens' DescribeObjectResponse (Maybe Text)

The content type of the object.

describeObjectResponse_contentLength :: Lens' DescribeObjectResponse (Maybe Natural)

The length of the object in bytes.

describeObjectResponse_cacheControl :: Lens' DescribeObjectResponse (Maybe Text)

An optional CacheControl header that allows the caller to control the object's cache behavior. Headers can be passed in as specified in the HTTP at https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.9.

Headers with a custom user-defined value are also accepted.

describeObjectResponse_lastModified :: Lens' DescribeObjectResponse (Maybe UTCTime)

The date and time that the object was last modified.

describeObjectResponse_httpStatus :: Lens' DescribeObjectResponse Int

The response's http status code.

PutObject

putObject_contentType :: Lens' PutObject (Maybe Text)

The content type of the object.

putObject_storageClass :: Lens' PutObject (Maybe StorageClass)

Indicates the storage class of a Put request. Defaults to high-performance temporal storage class, and objects are persisted into durable storage shortly after being received.

putObject_cacheControl :: Lens' PutObject (Maybe Text)

An optional CacheControl header that allows the caller to control the object's cache behavior. Headers can be passed in as specified in the HTTP at https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.9.

Headers with a custom user-defined value are also accepted.

putObject_uploadAvailability :: Lens' PutObject (Maybe UploadAvailability)

Indicates the availability of an object while it is still uploading. If the value is set to streaming, the object is available for downloading after some initial buffering but before the object is uploaded completely. If the value is set to standard, the object is available for downloading only when it is uploaded completely. The default value for this header is standard.

To use this header, you must also set the HTTP Transfer-Encoding header to chunked.

putObject_path :: Lens' PutObject Text

The path (including the file name) where the object is stored in the container. Format: <folder name>/<folder name>/<file name>

For example, to upload the file mlaw.avi to the folder path premium\canada in the container movies, enter the path premium/canada/mlaw.avi.

Do not include the container name in this path.

If the path includes any folders that don't exist yet, the service creates them. For example, suppose you have an existing premium/usa subfolder. If you specify premium/canada, the service creates a canada subfolder in the premium folder. You then have two subfolders, usa and canada, in the premium folder.

There is no correlation between the path to the source and the path (folders) in the container in AWS Elemental MediaStore.

For more information about folders and how they exist in a container, see the AWS Elemental MediaStore User Guide.

The file name is the name that is assigned to the file that you upload. The file can have the same name inside and outside of AWS Elemental MediaStore, or it can have the same name. The file name can include or omit an extension.

putObject_body :: Lens' PutObject HashedBody

The bytes to be stored.

putObjectResponse_eTag :: Lens' PutObjectResponse (Maybe Text)

Unique identifier of the object in the container.

putObjectResponse_contentSHA256 :: Lens' PutObjectResponse (Maybe Text)

The SHA256 digest of the object that is persisted.

putObjectResponse_storageClass :: Lens' PutObjectResponse (Maybe StorageClass)

The storage class where the object was persisted. The class should be “Temporal”.

putObjectResponse_httpStatus :: Lens' PutObjectResponse Int

The response's http status code.

GetObject

getObject_range :: Lens' GetObject (Maybe Text)

The range bytes of an object to retrieve. For more information about the Range header, see http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.35. AWS Elemental MediaStore ignores this header for partially uploaded objects that have streaming upload availability.

getObject_path :: Lens' GetObject Text

The path (including the file name) where the object is stored in the container. Format: <folder name>/<folder name>/<file name>

For example, to upload the file mlaw.avi to the folder path premium\canada in the container movies, enter the path premium/canada/mlaw.avi.

Do not include the container name in this path.

If the path includes any folders that don't exist yet, the service creates them. For example, suppose you have an existing premium/usa subfolder. If you specify premium/canada, the service creates a canada subfolder in the premium folder. You then have two subfolders, usa and canada, in the premium folder.

There is no correlation between the path to the source and the path (folders) in the container in AWS Elemental MediaStore.

For more information about folders and how they exist in a container, see the AWS Elemental MediaStore User Guide.

The file name is the name that is assigned to the file that you upload. The file can have the same name inside and outside of AWS Elemental MediaStore, or it can have the same name. The file name can include or omit an extension.

getObjectResponse_eTag :: Lens' GetObjectResponse (Maybe Text)

The ETag that represents a unique instance of the object.

getObjectResponse_contentType :: Lens' GetObjectResponse (Maybe Text)

The content type of the object.

getObjectResponse_contentRange :: Lens' GetObjectResponse (Maybe Text)

The range of bytes to retrieve.

getObjectResponse_contentLength :: Lens' GetObjectResponse (Maybe Natural)

The length of the object in bytes.

getObjectResponse_cacheControl :: Lens' GetObjectResponse (Maybe Text)

An optional CacheControl header that allows the caller to control the object's cache behavior. Headers can be passed in as specified in the HTTP spec at https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.9.

Headers with a custom user-defined value are also accepted.

getObjectResponse_lastModified :: Lens' GetObjectResponse (Maybe UTCTime)

The date and time that the object was last modified.

getObjectResponse_statusCode :: Lens' GetObjectResponse Int

The HTML status code of the request. Status codes ranging from 200 to 299 indicate success. All other status codes indicate the type of error that occurred.

getObjectResponse_body :: Lens' GetObjectResponse ResponseBody

The bytes of the object.

Types

Item

item_eTag :: Lens' Item (Maybe Text)

The ETag that represents a unique instance of the item.

item_contentType :: Lens' Item (Maybe Text)

The content type of the item.

item_contentLength :: Lens' Item (Maybe Natural)

The length of the item in bytes.

item_name :: Lens' Item (Maybe Text)

The name of the item.

item_lastModified :: Lens' Item (Maybe UTCTime)

The date and time that the item was last modified.

item_type :: Lens' Item (Maybe ItemType)

The item type (folder or object).