Network.AWS.MediaStoreData.PutObject

Description

Uploads an object to the specified path. Object sizes are limited to 25 MB for standard upload availability and 10 MB for streaming upload availability.

Creating a Request

data PutObject

See: newPutObject smart constructor.

Constructors

PutObject'

Fields

storageClass :: Maybe StorageClass

Indicates the storage class of a Put request. Defaults to high-performance temporal storage class, and objects are persisted into durable storage shortly after being received.

uploadAvailability :: Maybe UploadAvailability

Indicates the availability of an object while it is still uploading. If the value is set to streaming, the object is available for downloading after some initial buffering but before the object is uploaded completely. If the value is set to standard, the object is available for downloading only when it is uploaded completely. The default value for this header is standard.

To use this header, you must also set the HTTP Transfer-Encoding header to chunked.

cacheControl :: Maybe Text

An optional CacheControl header that allows the caller to control the object's cache behavior. Headers can be passed in as specified in the HTTP at https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.9.

Headers with a custom user-defined value are also accepted.

contentType :: Maybe Text

The content type of the object.

path :: Text

The path (including the file name) where the object is stored in the container. Format: <folder name>/<folder name>/<file name>

For example, to upload the file mlaw.avi to the folder path premium\canada in the container movies, enter the path premium/canada/mlaw.avi.

Do not include the container name in this path.

If the path includes any folders that don't exist yet, the service creates them. For example, suppose you have an existing premium/usa subfolder. If you specify premium/canada, the service creates a canada subfolder in the premium folder. You then have two subfolders, usa and canada, in the premium folder.

There is no correlation between the path to the source and the path (folders) in the container in AWS Elemental MediaStore.

For more information about folders and how they exist in a container, see the AWS Elemental MediaStore User Guide.

The file name is the name that is assigned to the file that you upload. The file can have the same name inside and outside of AWS Elemental MediaStore, or it can have the same name. The file name can include or omit an extension.

body :: HashedBody

The bytes to be stored.

newPutObject

Arguments

:: Text
$sel:path:PutObject'

-> HashedBody
$sel:body:PutObject'

-> PutObject

Create a value of PutObject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:storageClass:PutObject', putObject_storageClass - Indicates the storage class of a Put request. Defaults to high-performance temporal storage class, and objects are persisted into durable storage shortly after being received.

$sel:uploadAvailability:PutObject', putObject_uploadAvailability - Indicates the availability of an object while it is still uploading. If the value is set to streaming, the object is available for downloading after some initial buffering but before the object is uploaded completely. If the value is set to standard, the object is available for downloading only when it is uploaded completely. The default value for this header is standard.

To use this header, you must also set the HTTP Transfer-Encoding header to chunked.

$sel:cacheControl:PutObject', putObject_cacheControl - An optional CacheControl header that allows the caller to control the object's cache behavior. Headers can be passed in as specified in the HTTP at https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.9.

Headers with a custom user-defined value are also accepted.

$sel:contentType:PutObject', putObject_contentType - The content type of the object.

$sel:path:PutObject', putObject_path - The path (including the file name) where the object is stored in the container. Format: <folder name>/<folder name>/<file name>

For example, to upload the file mlaw.avi to the folder path premium\canada in the container movies, enter the path premium/canada/mlaw.avi.

Do not include the container name in this path.

If the path includes any folders that don't exist yet, the service creates them. For example, suppose you have an existing premium/usa subfolder. If you specify premium/canada, the service creates a canada subfolder in the premium folder. You then have two subfolders, usa and canada, in the premium folder.

There is no correlation between the path to the source and the path (folders) in the container in AWS Elemental MediaStore.

For more information about folders and how they exist in a container, see the AWS Elemental MediaStore User Guide.

The file name is the name that is assigned to the file that you upload. The file can have the same name inside and outside of AWS Elemental MediaStore, or it can have the same name. The file name can include or omit an extension.

$sel:body:PutObject', putObject_body - The bytes to be stored.

Request Lenses

putObject_storageClass :: Lens' PutObject (Maybe StorageClass)

Indicates the storage class of a Put request. Defaults to high-performance temporal storage class, and objects are persisted into durable storage shortly after being received.

putObject_uploadAvailability :: Lens' PutObject (Maybe UploadAvailability)

Indicates the availability of an object while it is still uploading. If the value is set to streaming, the object is available for downloading after some initial buffering but before the object is uploaded completely. If the value is set to standard, the object is available for downloading only when it is uploaded completely. The default value for this header is standard.

To use this header, you must also set the HTTP Transfer-Encoding header to chunked.

putObject_cacheControl :: Lens' PutObject (Maybe Text)

An optional CacheControl header that allows the caller to control the object's cache behavior. Headers can be passed in as specified in the HTTP at https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.9.

Headers with a custom user-defined value are also accepted.

putObject_contentType :: Lens' PutObject (Maybe Text)

The content type of the object.

putObject_path :: Lens' PutObject Text

The path (including the file name) where the object is stored in the container. Format: <folder name>/<folder name>/<file name>

For example, to upload the file mlaw.avi to the folder path premium\canada in the container movies, enter the path premium/canada/mlaw.avi.

Do not include the container name in this path.

If the path includes any folders that don't exist yet, the service creates them. For example, suppose you have an existing premium/usa subfolder. If you specify premium/canada, the service creates a canada subfolder in the premium folder. You then have two subfolders, usa and canada, in the premium folder.

There is no correlation between the path to the source and the path (folders) in the container in AWS Elemental MediaStore.

For more information about folders and how they exist in a container, see the AWS Elemental MediaStore User Guide.

The file name is the name that is assigned to the file that you upload. The file can have the same name inside and outside of AWS Elemental MediaStore, or it can have the same name. The file name can include or omit an extension.

putObject_body :: Lens' PutObject HashedBody

The bytes to be stored.

Destructuring the Response

data PutObjectResponse

See: newPutObjectResponse smart constructor.

Constructors

PutObjectResponse'

Fields

eTag :: Maybe Text

Unique identifier of the object in the container.

storageClass :: Maybe StorageClass

The storage class where the object was persisted. The class should be “Temporal”.

contentSHA256 :: Maybe Text

The SHA256 digest of the object that is persisted.

httpStatus :: Int

The response's http status code.

newPutObjectResponse

Arguments

:: Int
$sel:httpStatus:PutObjectResponse'

-> PutObjectResponse

Create a value of PutObjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:PutObjectResponse', putObjectResponse_eTag - Unique identifier of the object in the container.

$sel:storageClass:PutObject', putObjectResponse_storageClass - The storage class where the object was persisted. The class should be “Temporal”.

$sel:contentSHA256:PutObjectResponse', putObjectResponse_contentSHA256 - The SHA256 digest of the object that is persisted.

$sel:httpStatus:PutObjectResponse', putObjectResponse_httpStatus - The response's http status code.

Response Lenses

putObjectResponse_eTag :: Lens' PutObjectResponse (Maybe Text)

Unique identifier of the object in the container.

putObjectResponse_storageClass :: Lens' PutObjectResponse (Maybe StorageClass)

The storage class where the object was persisted. The class should be “Temporal”.

putObjectResponse_contentSHA256 :: Lens' PutObjectResponse (Maybe Text)

The SHA256 digest of the object that is persisted.

putObjectResponse_httpStatus :: Lens' PutObjectResponse Int

The response's http status code.