Network.AWS.MediaStoreData.Types

Description

Service Configuration

defaultService :: Service

API version 2017-09-01 of the Amazon Elemental MediaStore Data Plane SDK configuration.

Errors

_RequestedRangeNotSatisfiableException :: AsError a => Getting (First ServiceError) a ServiceError

The requested content range is not valid.

_ObjectNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Could not perform an operation on an object that does not exist.

_InternalServerError :: AsError a => Getting (First ServiceError) a ServiceError

The service is temporarily unavailable.

_ContainerNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified container was not found for the specified account.

ItemType

newtype ItemType

Constructors

ItemType'

Fields

fromItemType :: Text

Bundled Patterns

pattern ItemType_FOLDER :: ItemType

pattern ItemType_OBJECT :: ItemType

StorageClass

newtype StorageClass

Constructors

StorageClass'

Fields

fromStorageClass :: Text

Bundled Patterns

pattern StorageClass_TEMPORAL :: StorageClass

UploadAvailability

newtype UploadAvailability

Constructors

UploadAvailability'

Fields

fromUploadAvailability :: Text

Bundled Patterns

pattern UploadAvailability_STANDARD :: UploadAvailability

pattern UploadAvailability_STREAMING :: UploadAvailability

Item

data Item

A metadata entry for a folder or object.

See: newItem smart constructor.

Constructors

Item'

Fields

eTag :: Maybe Text

The ETag that represents a unique instance of the item.

contentType :: Maybe Text

The content type of the item.

contentLength :: Maybe Natural

The length of the item in bytes.

name :: Maybe Text

The name of the item.

lastModified :: Maybe POSIX

The date and time that the item was last modified.

type' :: Maybe ItemType

The item type (folder or object).

newItem :: Item

Create a value of Item with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eTag:Item', item_eTag - The ETag that represents a unique instance of the item.

$sel:contentType:Item', item_contentType - The content type of the item.

$sel:contentLength:Item', item_contentLength - The length of the item in bytes.

$sel:name:Item', item_name - The name of the item.

$sel:lastModified:Item', item_lastModified - The date and time that the item was last modified.

$sel:type':Item', item_type - The item type (folder or object).

item_eTag :: Lens' Item (Maybe Text)

The ETag that represents a unique instance of the item.

item_contentType :: Lens' Item (Maybe Text)

The content type of the item.

item_contentLength :: Lens' Item (Maybe Natural)

The length of the item in bytes.

item_name :: Lens' Item (Maybe Text)

The name of the item.

item_lastModified :: Lens' Item (Maybe UTCTime)

The date and time that the item was last modified.

item_type :: Lens' Item (Maybe ItemType)

The item type (folder or object).