Network.AWS.MediaTailor.CreateProgram

Description

Creates a program.

Creating a Request

data CreateProgram

See: newCreateProgram smart constructor.

Constructors

CreateProgram'

Fields

adBreaks :: Maybe [AdBreak]

The ad break configuration settings.

channelName :: Text

The identifier for the channel you are working on.

programName :: Text

The identifier for the program you are working on.

vodSourceName :: Text

The name that's used to refer to a VOD source.

scheduleConfiguration :: ScheduleConfiguration

The schedule configuration settings.

sourceLocationName :: Text

The name of the source location.

newCreateProgram

Arguments

:: Text
$sel:channelName:CreateProgram'

-> Text
$sel:programName:CreateProgram'

-> Text
$sel:vodSourceName:CreateProgram'

-> ScheduleConfiguration
$sel:scheduleConfiguration:CreateProgram'

-> Text
$sel:sourceLocationName:CreateProgram'

-> CreateProgram

Create a value of CreateProgram with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:adBreaks:CreateProgram', createProgram_adBreaks - The ad break configuration settings.

$sel:channelName:CreateProgram', createProgram_channelName - The identifier for the channel you are working on.

$sel:programName:CreateProgram', createProgram_programName - The identifier for the program you are working on.

$sel:vodSourceName:CreateProgram', createProgram_vodSourceName - The name that's used to refer to a VOD source.

$sel:scheduleConfiguration:CreateProgram', createProgram_scheduleConfiguration - The schedule configuration settings.

$sel:sourceLocationName:CreateProgram', createProgram_sourceLocationName - The name of the source location.

Request Lenses

createProgram_adBreaks :: Lens' CreateProgram (Maybe [AdBreak])

The ad break configuration settings.

createProgram_channelName :: Lens' CreateProgram Text

The identifier for the channel you are working on.

createProgram_programName :: Lens' CreateProgram Text

The identifier for the program you are working on.

createProgram_vodSourceName :: Lens' CreateProgram Text

The name that's used to refer to a VOD source.

createProgram_scheduleConfiguration :: Lens' CreateProgram ScheduleConfiguration

The schedule configuration settings.

createProgram_sourceLocationName :: Lens' CreateProgram Text

The name of the source location.

Destructuring the Response

data CreateProgramResponse

See: newCreateProgramResponse smart constructor.

Constructors

CreateProgramResponse'

Fields

creationTime :: Maybe POSIX

The timestamp of when the program was created.

sourceLocationName :: Maybe Text

The source location name.

arn :: Maybe Text

The ARN of the program.

programName :: Maybe Text

The name of the program.

adBreaks :: Maybe [AdBreak]

The ad break configuration settings.

channelName :: Maybe Text

The name of the channel that the program belongs to.

scheduledStartTime :: Maybe POSIX

The date and time that the program is scheduled to start in ISO 8601 format and Coordinated Universal Time (UTC). For example, the value 2021-03-27T17:48:16.751Z represents March 27, 2021 at 17:48:16.751 UTC.

vodSourceName :: Maybe Text

The name that's used to refer to a VOD source.

httpStatus :: Int

The response's http status code.

newCreateProgramResponse

Arguments

:: Int
$sel:httpStatus:CreateProgramResponse'

-> CreateProgramResponse

Create a value of CreateProgramResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:CreateProgramResponse', createProgramResponse_creationTime - The timestamp of when the program was created.

$sel:sourceLocationName:CreateProgram', createProgramResponse_sourceLocationName - The source location name.

$sel:arn:CreateProgramResponse', createProgramResponse_arn - The ARN of the program.

$sel:programName:CreateProgram', createProgramResponse_programName - The name of the program.

$sel:adBreaks:CreateProgram', createProgramResponse_adBreaks - The ad break configuration settings.

$sel:channelName:CreateProgram', createProgramResponse_channelName - The name of the channel that the program belongs to.

$sel:scheduledStartTime:CreateProgramResponse', createProgramResponse_scheduledStartTime - The date and time that the program is scheduled to start in ISO 8601 format and Coordinated Universal Time (UTC). For example, the value 2021-03-27T17:48:16.751Z represents March 27, 2021 at 17:48:16.751 UTC.

$sel:vodSourceName:CreateProgram', createProgramResponse_vodSourceName - The name that's used to refer to a VOD source.

$sel:httpStatus:CreateProgramResponse', createProgramResponse_httpStatus - The response's http status code.

Response Lenses

createProgramResponse_creationTime :: Lens' CreateProgramResponse (Maybe UTCTime)

The timestamp of when the program was created.

createProgramResponse_sourceLocationName :: Lens' CreateProgramResponse (Maybe Text)

The source location name.

createProgramResponse_arn :: Lens' CreateProgramResponse (Maybe Text)

The ARN of the program.

createProgramResponse_programName :: Lens' CreateProgramResponse (Maybe Text)

The name of the program.

createProgramResponse_adBreaks :: Lens' CreateProgramResponse (Maybe [AdBreak])

The ad break configuration settings.

createProgramResponse_channelName :: Lens' CreateProgramResponse (Maybe Text)

The name of the channel that the program belongs to.

createProgramResponse_scheduledStartTime :: Lens' CreateProgramResponse (Maybe UTCTime)

The date and time that the program is scheduled to start in ISO 8601 format and Coordinated Universal Time (UTC). For example, the value 2021-03-27T17:48:16.751Z represents March 27, 2021 at 17:48:16.751 UTC.

createProgramResponse_vodSourceName :: Lens' CreateProgramResponse (Maybe Text)

The name that's used to refer to a VOD source.

createProgramResponse_httpStatus :: Lens' CreateProgramResponse Int

The response's http status code.