Network.AWS.MediaTailor.DescribeProgram

Description

Retrieves the properties of the requested program.

Creating a Request

data DescribeProgram

See: newDescribeProgram smart constructor.

Constructors

DescribeProgram'

Fields

channelName :: Text

The identifier for the channel you are working on.

programName :: Text

The identifier for the program you are working on.

newDescribeProgram

Arguments

:: Text
$sel:channelName:DescribeProgram'

-> Text
$sel:programName:DescribeProgram'

-> DescribeProgram

Create a value of DescribeProgram with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelName:DescribeProgram', describeProgram_channelName - The identifier for the channel you are working on.

$sel:programName:DescribeProgram', describeProgram_programName - The identifier for the program you are working on.

Request Lenses

describeProgram_channelName :: Lens' DescribeProgram Text

The identifier for the channel you are working on.

describeProgram_programName :: Lens' DescribeProgram Text

The identifier for the program you are working on.

Destructuring the Response

data DescribeProgramResponse

See: newDescribeProgramResponse smart constructor.

Constructors

DescribeProgramResponse'

Fields

creationTime :: Maybe POSIX

The timestamp of when the program was created.

sourceLocationName :: Maybe Text

The source location name.

arn :: Maybe Text

The ARN of the program.

programName :: Maybe Text

The name of the program.

adBreaks :: Maybe [AdBreak]

The ad break configuration settings.

channelName :: Maybe Text

The name of the channel that the program belongs to.

scheduledStartTime :: Maybe POSIX

The date and time that the program is scheduled to start in ISO 8601 format and Coordinated Universal Time (UTC). For example, the value 2021-03-27T17:48:16.751Z represents March 27, 2021 at 17:48:16.751 UTC.

vodSourceName :: Maybe Text

The name that's used to refer to a VOD source.

httpStatus :: Int

The response's http status code.

newDescribeProgramResponse

Arguments

:: Int
$sel:httpStatus:DescribeProgramResponse'

-> DescribeProgramResponse

Create a value of DescribeProgramResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeProgramResponse', describeProgramResponse_creationTime - The timestamp of when the program was created.

$sel:sourceLocationName:DescribeProgramResponse', describeProgramResponse_sourceLocationName - The source location name.

$sel:arn:DescribeProgramResponse', describeProgramResponse_arn - The ARN of the program.

$sel:programName:DescribeProgram', describeProgramResponse_programName - The name of the program.

$sel:adBreaks:DescribeProgramResponse', describeProgramResponse_adBreaks - The ad break configuration settings.

$sel:channelName:DescribeProgram', describeProgramResponse_channelName - The name of the channel that the program belongs to.

$sel:scheduledStartTime:DescribeProgramResponse', describeProgramResponse_scheduledStartTime - The date and time that the program is scheduled to start in ISO 8601 format and Coordinated Universal Time (UTC). For example, the value 2021-03-27T17:48:16.751Z represents March 27, 2021 at 17:48:16.751 UTC.

$sel:vodSourceName:DescribeProgramResponse', describeProgramResponse_vodSourceName - The name that's used to refer to a VOD source.

$sel:httpStatus:DescribeProgramResponse', describeProgramResponse_httpStatus - The response's http status code.

Response Lenses

describeProgramResponse_creationTime :: Lens' DescribeProgramResponse (Maybe UTCTime)

The timestamp of when the program was created.

describeProgramResponse_sourceLocationName :: Lens' DescribeProgramResponse (Maybe Text)

The source location name.

describeProgramResponse_arn :: Lens' DescribeProgramResponse (Maybe Text)

The ARN of the program.

describeProgramResponse_programName :: Lens' DescribeProgramResponse (Maybe Text)

The name of the program.

describeProgramResponse_adBreaks :: Lens' DescribeProgramResponse (Maybe [AdBreak])

The ad break configuration settings.

describeProgramResponse_channelName :: Lens' DescribeProgramResponse (Maybe Text)

The name of the channel that the program belongs to.

describeProgramResponse_scheduledStartTime :: Lens' DescribeProgramResponse (Maybe UTCTime)

The date and time that the program is scheduled to start in ISO 8601 format and Coordinated Universal Time (UTC). For example, the value 2021-03-27T17:48:16.751Z represents March 27, 2021 at 17:48:16.751 UTC.

describeProgramResponse_vodSourceName :: Lens' DescribeProgramResponse (Maybe Text)

The name that's used to refer to a VOD source.

describeProgramResponse_httpStatus :: Lens' DescribeProgramResponse Int

The response's http status code.