Network.AWS.MediaTailor.DescribeVodSource

Description

Provides details about a specific VOD source in a specific source location.

Creating a Request

data DescribeVodSource

See: newDescribeVodSource smart constructor.

Constructors

DescribeVodSource'

Fields

sourceLocationName :: Text

The identifier for the source location you are working on.

vodSourceName :: Text

The identifier for the VOD source you are working on.

newDescribeVodSource

Arguments

:: Text
$sel:sourceLocationName:DescribeVodSource'

-> Text
$sel:vodSourceName:DescribeVodSource'

-> DescribeVodSource

Create a value of DescribeVodSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceLocationName:DescribeVodSource', describeVodSource_sourceLocationName - The identifier for the source location you are working on.

$sel:vodSourceName:DescribeVodSource', describeVodSource_vodSourceName - The identifier for the VOD source you are working on.

Request Lenses

describeVodSource_sourceLocationName :: Lens' DescribeVodSource Text

The identifier for the source location you are working on.

describeVodSource_vodSourceName :: Lens' DescribeVodSource Text

The identifier for the VOD source you are working on.

Destructuring the Response

data DescribeVodSourceResponse

See: newDescribeVodSourceResponse smart constructor.

Constructors

DescribeVodSourceResponse'

Fields

creationTime :: Maybe POSIX

The timestamp that indicates when the VOD source was created.

sourceLocationName :: Maybe Text

The name of the source location associated with the VOD source.

arn :: Maybe Text

The ARN of the VOD source.

lastModifiedTime :: Maybe POSIX

The ARN for the VOD source.

httpPackageConfigurations :: Maybe [HttpPackageConfiguration]

The HTTP package configurations.

vodSourceName :: Maybe Text

The name of the VOD source.

tags :: Maybe (HashMap Text Text)

The tags assigned to the VOD source.

httpStatus :: Int

The response's http status code.

newDescribeVodSourceResponse

Arguments

:: Int
$sel:httpStatus:DescribeVodSourceResponse'

-> DescribeVodSourceResponse

Create a value of DescribeVodSourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:DescribeVodSourceResponse', describeVodSourceResponse_creationTime - The timestamp that indicates when the VOD source was created.

$sel:sourceLocationName:DescribeVodSource', describeVodSourceResponse_sourceLocationName - The name of the source location associated with the VOD source.

$sel:arn:DescribeVodSourceResponse', describeVodSourceResponse_arn - The ARN of the VOD source.

$sel:lastModifiedTime:DescribeVodSourceResponse', describeVodSourceResponse_lastModifiedTime - The ARN for the VOD source.

$sel:httpPackageConfigurations:DescribeVodSourceResponse', describeVodSourceResponse_httpPackageConfigurations - The HTTP package configurations.

$sel:vodSourceName:DescribeVodSource', describeVodSourceResponse_vodSourceName - The name of the VOD source.

$sel:tags:DescribeVodSourceResponse', describeVodSourceResponse_tags - The tags assigned to the VOD source.

$sel:httpStatus:DescribeVodSourceResponse', describeVodSourceResponse_httpStatus - The response's http status code.

Response Lenses

describeVodSourceResponse_creationTime :: Lens' DescribeVodSourceResponse (Maybe UTCTime)

The timestamp that indicates when the VOD source was created.

describeVodSourceResponse_sourceLocationName :: Lens' DescribeVodSourceResponse (Maybe Text)

The name of the source location associated with the VOD source.

describeVodSourceResponse_arn :: Lens' DescribeVodSourceResponse (Maybe Text)

The ARN of the VOD source.

describeVodSourceResponse_lastModifiedTime :: Lens' DescribeVodSourceResponse (Maybe UTCTime)

The ARN for the VOD source.

describeVodSourceResponse_httpPackageConfigurations :: Lens' DescribeVodSourceResponse (Maybe [HttpPackageConfiguration])

The HTTP package configurations.

describeVodSourceResponse_vodSourceName :: Lens' DescribeVodSourceResponse (Maybe Text)

The name of the VOD source.

describeVodSourceResponse_tags :: Lens' DescribeVodSourceResponse (Maybe (HashMap Text Text))

The tags assigned to the VOD source.

describeVodSourceResponse_httpStatus :: Lens' DescribeVodSourceResponse Int

The response's http status code.