Network.AWS.MediaTailor.GetChannelSchedule

Description

Retrieves information about your channel's schedule.

This operation returns paginated results.

Creating a Request

data GetChannelSchedule

See: newGetChannelSchedule smart constructor.

Constructors

GetChannelSchedule'

Fields

nextToken :: Maybe Text

Pagination token from the GET list request. Use the token to fetch the next page of results.

durationMinutes :: Maybe Text

The schedule duration in minutes. The maximum duration is 4320 minutes (three days).

maxResults :: Maybe Natural

Upper bound on number of records to return. The maximum number of results is 100.

channelName :: Text

The identifier for the channel you are working on.

newGetChannelSchedule

Arguments

:: Text
$sel:channelName:GetChannelSchedule'

-> GetChannelSchedule

Create a value of GetChannelSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetChannelSchedule', getChannelSchedule_nextToken - Pagination token from the GET list request. Use the token to fetch the next page of results.

$sel:durationMinutes:GetChannelSchedule', getChannelSchedule_durationMinutes - The schedule duration in minutes. The maximum duration is 4320 minutes (three days).

$sel:maxResults:GetChannelSchedule', getChannelSchedule_maxResults - Upper bound on number of records to return. The maximum number of results is 100.

$sel:channelName:GetChannelSchedule', getChannelSchedule_channelName - The identifier for the channel you are working on.

Request Lenses

getChannelSchedule_nextToken :: Lens' GetChannelSchedule (Maybe Text)

Pagination token from the GET list request. Use the token to fetch the next page of results.

getChannelSchedule_durationMinutes :: Lens' GetChannelSchedule (Maybe Text)

The schedule duration in minutes. The maximum duration is 4320 minutes (three days).

getChannelSchedule_maxResults :: Lens' GetChannelSchedule (Maybe Natural)

Upper bound on number of records to return. The maximum number of results is 100.

getChannelSchedule_channelName :: Lens' GetChannelSchedule Text

The identifier for the channel you are working on.

Destructuring the Response

data GetChannelScheduleResponse

See: newGetChannelScheduleResponse smart constructor.

Constructors

GetChannelScheduleResponse'

Fields

items :: Maybe [ScheduleEntry]

An array of schedule entries for the channel.

nextToken :: Maybe Text

Pagination token from the GET list request. Use the token to fetch the next page of results.

httpStatus :: Int

The response's http status code.

newGetChannelScheduleResponse

Arguments

:: Int
$sel:httpStatus:GetChannelScheduleResponse'

-> GetChannelScheduleResponse

Create a value of GetChannelScheduleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:items:GetChannelScheduleResponse', getChannelScheduleResponse_items - An array of schedule entries for the channel.

$sel:nextToken:GetChannelSchedule', getChannelScheduleResponse_nextToken - Pagination token from the GET list request. Use the token to fetch the next page of results.

$sel:httpStatus:GetChannelScheduleResponse', getChannelScheduleResponse_httpStatus - The response's http status code.

Response Lenses

getChannelScheduleResponse_items :: Lens' GetChannelScheduleResponse (Maybe [ScheduleEntry])

An array of schedule entries for the channel.

getChannelScheduleResponse_nextToken :: Lens' GetChannelScheduleResponse (Maybe Text)

Pagination token from the GET list request. Use the token to fetch the next page of results.

getChannelScheduleResponse_httpStatus :: Lens' GetChannelScheduleResponse Int

The response's http status code.