Network.AWS.MediaTailor.Types.AdBreak

Description

Documentation

data AdBreak

Ad break configuration parameters.

See: newAdBreak smart constructor.

Constructors

AdBreak'

Fields

spliceInsertMessage :: Maybe SpliceInsertMessage

This defines the SCTE-35 splice_insert() message inserted around the ad. For information about using splice_insert(), see the SCTE-35 specficiaiton, section 9.7.3.1.

messageType :: Maybe MessageType

The SCTE-35 ad insertion type. Accepted value: SPLICE_INSERT.

slate :: Maybe SlateSource

Ad break slate configuration.

offsetMillis :: Maybe Integer

How long (in milliseconds) after the beginning of the program that an ad starts. This value must fall within 100ms of a segment boundary, otherwise the ad break will be skipped.

newAdBreak :: AdBreak

Create a value of AdBreak with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:spliceInsertMessage:AdBreak', adBreak_spliceInsertMessage - This defines the SCTE-35 splice_insert() message inserted around the ad. For information about using splice_insert(), see the SCTE-35 specficiaiton, section 9.7.3.1.

$sel:messageType:AdBreak', adBreak_messageType - The SCTE-35 ad insertion type. Accepted value: SPLICE_INSERT.

$sel:slate:AdBreak', adBreak_slate - Ad break slate configuration.

$sel:offsetMillis:AdBreak', adBreak_offsetMillis - How long (in milliseconds) after the beginning of the program that an ad starts. This value must fall within 100ms of a segment boundary, otherwise the ad break will be skipped.

adBreak_spliceInsertMessage :: Lens' AdBreak (Maybe SpliceInsertMessage)

This defines the SCTE-35 splice_insert() message inserted around the ad. For information about using splice_insert(), see the SCTE-35 specficiaiton, section 9.7.3.1.

adBreak_messageType :: Lens' AdBreak (Maybe MessageType)

The SCTE-35 ad insertion type. Accepted value: SPLICE_INSERT.

adBreak_slate :: Lens' AdBreak (Maybe SlateSource)

Ad break slate configuration.

adBreak_offsetMillis :: Lens' AdBreak (Maybe Integer)

How long (in milliseconds) after the beginning of the program that an ad starts. This value must fall within 100ms of a segment boundary, otherwise the ad break will be skipped.