Network.AWS.MediaTailor.Types.ManifestProcessingRules

Description

Documentation

data ManifestProcessingRules

The configuration for manifest processing rules. Manifest processing rules enable customization of the personalized manifests created by MediaTailor.

See: newManifestProcessingRules smart constructor.

Constructors

ManifestProcessingRules'

Fields

adMarkerPassthrough :: Maybe AdMarkerPassthrough

For HLS, when set to true, MediaTailor passes through EXT-X-CUE-IN, EXT-X-CUE-OUT, and EXT-X-SPLICEPOINT-SCTE35 ad markers from the origin manifest to the MediaTailor personalized manifest.

No logic is applied to these ad markers. For example, if EXT-X-CUE-OUT has a value of 60, but no ads are filled for that ad break, MediaTailor will not set the value to 0.

newManifestProcessingRules :: ManifestProcessingRules

Create a value of ManifestProcessingRules with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:adMarkerPassthrough:ManifestProcessingRules', manifestProcessingRules_adMarkerPassthrough - For HLS, when set to true, MediaTailor passes through EXT-X-CUE-IN, EXT-X-CUE-OUT, and EXT-X-SPLICEPOINT-SCTE35 ad markers from the origin manifest to the MediaTailor personalized manifest.

No logic is applied to these ad markers. For example, if EXT-X-CUE-OUT has a value of 60, but no ads are filled for that ad break, MediaTailor will not set the value to 0.

manifestProcessingRules_adMarkerPassthrough :: Lens' ManifestProcessingRules (Maybe AdMarkerPassthrough)

For HLS, when set to true, MediaTailor passes through EXT-X-CUE-IN, EXT-X-CUE-OUT, and EXT-X-SPLICEPOINT-SCTE35 ad markers from the origin manifest to the MediaTailor personalized manifest.

No logic is applied to these ad markers. For example, if EXT-X-CUE-OUT has a value of 60, but no ads are filled for that ad break, MediaTailor will not set the value to 0.