Network.AWS.MediaTailor.Types.PrefetchRetrieval

Description

Documentation

data PrefetchRetrieval

A complex type that contains settings governing when MediaTailor prefetches ads, and which dynamic variables that MediaTailor includes in the request to the ad decision server.

See: newPrefetchRetrieval smart constructor.

Constructors

PrefetchRetrieval'

Fields

startTime :: Maybe POSIX

The time when prefetch retrievals can start for this break. Ad prefetching will be attempted for manifest requests that occur at or after this time. Defaults to the current time. If not specified, the prefetch retrieval starts as soon as possible.

dynamicVariables :: Maybe (HashMap Text Text)

The dynamic variables to use for substitution during prefetch requests to the ad decision server (ADS).

You intially configure dynamic variables for the ADS URL when you set up your playback configuration. When you specify DynamicVariables for prefetch retrieval, MediaTailor includes the dynamic variables in the request to the ADS.

endTime :: POSIX

The time when prefetch retrieval ends for the ad break. Prefetching will be attempted for manifest requests that occur at or before this time.

newPrefetchRetrieval

Arguments

:: UTCTime
$sel:endTime:PrefetchRetrieval'

-> PrefetchRetrieval

Create a value of PrefetchRetrieval with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startTime:PrefetchRetrieval', prefetchRetrieval_startTime - The time when prefetch retrievals can start for this break. Ad prefetching will be attempted for manifest requests that occur at or after this time. Defaults to the current time. If not specified, the prefetch retrieval starts as soon as possible.

$sel:dynamicVariables:PrefetchRetrieval', prefetchRetrieval_dynamicVariables - The dynamic variables to use for substitution during prefetch requests to the ad decision server (ADS).

You intially configure dynamic variables for the ADS URL when you set up your playback configuration. When you specify DynamicVariables for prefetch retrieval, MediaTailor includes the dynamic variables in the request to the ADS.

$sel:endTime:PrefetchRetrieval', prefetchRetrieval_endTime - The time when prefetch retrieval ends for the ad break. Prefetching will be attempted for manifest requests that occur at or before this time.

prefetchRetrieval_startTime :: Lens' PrefetchRetrieval (Maybe UTCTime)

The time when prefetch retrievals can start for this break. Ad prefetching will be attempted for manifest requests that occur at or after this time. Defaults to the current time. If not specified, the prefetch retrieval starts as soon as possible.

prefetchRetrieval_dynamicVariables :: Lens' PrefetchRetrieval (Maybe (HashMap Text Text))

The dynamic variables to use for substitution during prefetch requests to the ad decision server (ADS).

You intially configure dynamic variables for the ADS URL when you set up your playback configuration. When you specify DynamicVariables for prefetch retrieval, MediaTailor includes the dynamic variables in the request to the ADS.

prefetchRetrieval_endTime :: Lens' PrefetchRetrieval UTCTime

The time when prefetch retrieval ends for the ad break. Prefetching will be attempted for manifest requests that occur at or before this time.