Network.AWS.MediaTailor.Types.ScheduleEntry

Description

Documentation

data ScheduleEntry

The properties for a schedule.

See: newScheduleEntry smart constructor.

Constructors

ScheduleEntry'

Fields

scheduleAdBreaks :: Maybe [ScheduleAdBreak]

The schedule's ad break properties.

approximateDurationSeconds :: Maybe Integer

The approximate duration of this program, in seconds.

approximateStartTime :: Maybe POSIX

The approximate time that the program will start playing.

scheduleEntryType :: Maybe ScheduleEntryType

The type of schedule entry.

Valid values: PROGRAM or FILLER_SLATE.

vodSourceName :: Text

The name of the VOD source.

channelName :: Text

The name of the channel that uses this schedule.

sourceLocationName :: Text

The name of the source location.

arn :: Text

The ARN of the program.

programName :: Text

The name of the program.

newScheduleEntry

Arguments

:: Text
$sel:vodSourceName:ScheduleEntry'

-> Text
$sel:channelName:ScheduleEntry'

-> Text
$sel:sourceLocationName:ScheduleEntry'

-> Text
$sel:arn:ScheduleEntry'

-> Text
$sel:programName:ScheduleEntry'

-> ScheduleEntry

Create a value of ScheduleEntry with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scheduleAdBreaks:ScheduleEntry', scheduleEntry_scheduleAdBreaks - The schedule's ad break properties.

$sel:approximateDurationSeconds:ScheduleEntry', scheduleEntry_approximateDurationSeconds - The approximate duration of this program, in seconds.

$sel:approximateStartTime:ScheduleEntry', scheduleEntry_approximateStartTime - The approximate time that the program will start playing.

$sel:scheduleEntryType:ScheduleEntry', scheduleEntry_scheduleEntryType - The type of schedule entry.

Valid values: PROGRAM or FILLER_SLATE.

$sel:vodSourceName:ScheduleEntry', scheduleEntry_vodSourceName - The name of the VOD source.

$sel:channelName:ScheduleEntry', scheduleEntry_channelName - The name of the channel that uses this schedule.

$sel:sourceLocationName:ScheduleEntry', scheduleEntry_sourceLocationName - The name of the source location.

$sel:arn:ScheduleEntry', scheduleEntry_arn - The ARN of the program.

$sel:programName:ScheduleEntry', scheduleEntry_programName - The name of the program.

scheduleEntry_scheduleAdBreaks :: Lens' ScheduleEntry (Maybe [ScheduleAdBreak])

The schedule's ad break properties.

scheduleEntry_approximateDurationSeconds :: Lens' ScheduleEntry (Maybe Integer)

The approximate duration of this program, in seconds.

scheduleEntry_approximateStartTime :: Lens' ScheduleEntry (Maybe UTCTime)

The approximate time that the program will start playing.

scheduleEntry_scheduleEntryType :: Lens' ScheduleEntry (Maybe ScheduleEntryType)

The type of schedule entry.

Valid values: PROGRAM or FILLER_SLATE.

scheduleEntry_vodSourceName :: Lens' ScheduleEntry Text

The name of the VOD source.

scheduleEntry_channelName :: Lens' ScheduleEntry Text

The name of the channel that uses this schedule.

scheduleEntry_sourceLocationName :: Lens' ScheduleEntry Text

The name of the source location.

scheduleEntry_arn :: Lens' ScheduleEntry Text

The ARN of the program.

scheduleEntry_programName :: Lens' ScheduleEntry Text

The name of the program.