Network.AWS.MediaTailor.Types.SecretsManagerAccessTokenConfiguration

Description

Documentation

data SecretsManagerAccessTokenConfiguration

AWS Secrets Manager access token configuration parameters. For information about Secrets Manager access token authentication, see Working with AWS Secrets Manager access token authentication.

See: newSecretsManagerAccessTokenConfiguration smart constructor.

Constructors

SecretsManagerAccessTokenConfiguration'

Fields

headerName :: Maybe Text

The name of the HTTP header used to supply the access token in requests to the source location.

secretStringKey :: Maybe Text

The AWS Secrets Manager SecretString key associated with the access token. MediaTailor uses the key to look up SecretString key and value pair containing the access token.

secretArn :: Maybe Text

The Amazon Resource Name (ARN) of the AWS Secrets Manager secret that contains the access token.

newSecretsManagerAccessTokenConfiguration :: SecretsManagerAccessTokenConfiguration

Create a value of SecretsManagerAccessTokenConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:headerName:SecretsManagerAccessTokenConfiguration', secretsManagerAccessTokenConfiguration_headerName - The name of the HTTP header used to supply the access token in requests to the source location.

$sel:secretStringKey:SecretsManagerAccessTokenConfiguration', secretsManagerAccessTokenConfiguration_secretStringKey - The AWS Secrets Manager SecretString key associated with the access token. MediaTailor uses the key to look up SecretString key and value pair containing the access token.

$sel:secretArn:SecretsManagerAccessTokenConfiguration', secretsManagerAccessTokenConfiguration_secretArn - The Amazon Resource Name (ARN) of the AWS Secrets Manager secret that contains the access token.

secretsManagerAccessTokenConfiguration_headerName :: Lens' SecretsManagerAccessTokenConfiguration (Maybe Text)

The name of the HTTP header used to supply the access token in requests to the source location.

secretsManagerAccessTokenConfiguration_secretStringKey :: Lens' SecretsManagerAccessTokenConfiguration (Maybe Text)

The AWS Secrets Manager SecretString key associated with the access token. MediaTailor uses the key to look up SecretString key and value pair containing the access token.

secretsManagerAccessTokenConfiguration_secretArn :: Lens' SecretsManagerAccessTokenConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the AWS Secrets Manager secret that contains the access token.