Network.AWS.MediaTailor.Types.Transition

Description

Documentation

data Transition

Program transition configuration.

See: newTransition smart constructor.

Constructors

Transition'

Fields

scheduledStartTimeMillis :: Maybe Integer

The date and time that the program is scheduled to start, in epoch milliseconds.

relativeProgram :: Maybe Text

The name of the program that this program will be inserted next to, as defined by RelativePosition.

type' :: Text

Defines when the program plays in the schedule. You can set the value to ABSOLUTE or RELATIVE.

ABSOLUTE - The program plays at a specific wall clock time. This setting can only be used for channels using the LINEAR PlaybackMode.

Note the following considerations when using ABSOLUTE transitions:

If the preceding program in the schedule has a duration that extends past the wall clock time, MediaTailor truncates the preceding program on a common segment boundary.

If there are gaps in playback, MediaTailor plays the FillerSlate you configured for your linear channel.

RELATIVE - The program is inserted into the schedule either before or after a program that you specify via RelativePosition.

relativePosition :: RelativePosition

The position where this program will be inserted relative to the RelativePosition.

newTransition

Arguments

:: Text
$sel:type':Transition'

-> RelativePosition
$sel:relativePosition:Transition'

-> Transition

Create a value of Transition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:scheduledStartTimeMillis:Transition', transition_scheduledStartTimeMillis - The date and time that the program is scheduled to start, in epoch milliseconds.

$sel:relativeProgram:Transition', transition_relativeProgram - The name of the program that this program will be inserted next to, as defined by RelativePosition.

$sel:type':Transition', transition_type - Defines when the program plays in the schedule. You can set the value to ABSOLUTE or RELATIVE.

ABSOLUTE - The program plays at a specific wall clock time. This setting can only be used for channels using the LINEAR PlaybackMode.

Note the following considerations when using ABSOLUTE transitions:

If the preceding program in the schedule has a duration that extends past the wall clock time, MediaTailor truncates the preceding program on a common segment boundary.

If there are gaps in playback, MediaTailor plays the FillerSlate you configured for your linear channel.

RELATIVE - The program is inserted into the schedule either before or after a program that you specify via RelativePosition.

$sel:relativePosition:Transition', transition_relativePosition - The position where this program will be inserted relative to the RelativePosition.

transition_scheduledStartTimeMillis :: Lens' Transition (Maybe Integer)

The date and time that the program is scheduled to start, in epoch milliseconds.

transition_relativeProgram :: Lens' Transition (Maybe Text)

The name of the program that this program will be inserted next to, as defined by RelativePosition.

transition_type :: Lens' Transition Text

Defines when the program plays in the schedule. You can set the value to ABSOLUTE or RELATIVE.

ABSOLUTE - The program plays at a specific wall clock time. This setting can only be used for channels using the LINEAR PlaybackMode.

Note the following considerations when using ABSOLUTE transitions:

If the preceding program in the schedule has a duration that extends past the wall clock time, MediaTailor truncates the preceding program on a common segment boundary.

If there are gaps in playback, MediaTailor plays the FillerSlate you configured for your linear channel.

RELATIVE - The program is inserted into the schedule either before or after a program that you specify via RelativePosition.

transition_relativePosition :: Lens' Transition RelativePosition

The position where this program will be inserted relative to the RelativePosition.