Network.AWS.MediaTailor.Types.VodSource

Description

Documentation

data VodSource

VOD source configuration parameters.

See: newVodSource smart constructor.

Constructors

VodSource'

Fields

creationTime :: Maybe POSIX

The timestamp that indicates when the VOD source was created.

lastModifiedTime :: Maybe POSIX

The timestamp that indicates when the VOD source was last modified.

tags :: Maybe (HashMap Text Text)

The tags assigned to the VOD source.

vodSourceName :: Text

The name of the VOD source.

sourceLocationName :: Text

The name of the source location that the VOD source is associated with.

httpPackageConfigurations :: [HttpPackageConfiguration]

The HTTP package configurations for the VOD source.

arn :: Text

The ARN for the VOD source.

newVodSource

Arguments

:: Text
$sel:vodSourceName:VodSource'

-> Text
$sel:sourceLocationName:VodSource'

-> Text
$sel:arn:VodSource'

-> VodSource

Create a value of VodSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:VodSource', vodSource_creationTime - The timestamp that indicates when the VOD source was created.

$sel:lastModifiedTime:VodSource', vodSource_lastModifiedTime - The timestamp that indicates when the VOD source was last modified.

$sel:tags:VodSource', vodSource_tags - The tags assigned to the VOD source.

$sel:vodSourceName:VodSource', vodSource_vodSourceName - The name of the VOD source.

$sel:sourceLocationName:VodSource', vodSource_sourceLocationName - The name of the source location that the VOD source is associated with.

$sel:httpPackageConfigurations:VodSource', vodSource_httpPackageConfigurations - The HTTP package configurations for the VOD source.

$sel:arn:VodSource', vodSource_arn - The ARN for the VOD source.

vodSource_creationTime :: Lens' VodSource (Maybe UTCTime)

The timestamp that indicates when the VOD source was created.

vodSource_lastModifiedTime :: Lens' VodSource (Maybe UTCTime)

The timestamp that indicates when the VOD source was last modified.

vodSource_tags :: Lens' VodSource (Maybe (HashMap Text Text))

The tags assigned to the VOD source.

vodSource_vodSourceName :: Lens' VodSource Text

The name of the VOD source.

vodSource_sourceLocationName :: Lens' VodSource Text

The name of the source location that the VOD source is associated with.

vodSource_httpPackageConfigurations :: Lens' VodSource [HttpPackageConfiguration]

The HTTP package configurations for the VOD source.

vodSource_arn :: Lens' VodSource Text

The ARN for the VOD source.