Network.AWS.MediaTailor.UpdateChannel

Description

Updates an existing channel.

Creating a Request

data UpdateChannel

See: newUpdateChannel smart constructor.

Constructors

UpdateChannel'

Fields

channelName :: Text

The identifier for the channel you are working on.

outputs :: [RequestOutputItem]

The channel's output properties.

newUpdateChannel

Arguments

:: Text
$sel:channelName:UpdateChannel'

-> UpdateChannel

Create a value of UpdateChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelName:UpdateChannel', updateChannel_channelName - The identifier for the channel you are working on.

$sel:outputs:UpdateChannel', updateChannel_outputs - The channel's output properties.

Request Lenses

updateChannel_channelName :: Lens' UpdateChannel Text

The identifier for the channel you are working on.

updateChannel_outputs :: Lens' UpdateChannel [RequestOutputItem]

The channel's output properties.

Destructuring the Response

data UpdateChannelResponse

See: newUpdateChannelResponse smart constructor.

Constructors

UpdateChannelResponse'

Fields

creationTime :: Maybe POSIX

The timestamp of when the channel was created.

arn :: Maybe Text

The ARN of the channel.

lastModifiedTime :: Maybe POSIX

The timestamp of when the channel was last modified.

playbackMode :: Maybe Text

The channel's playback mode.

channelName :: Maybe Text

The name of the channel.

outputs :: Maybe [ResponseOutputItem]

The channel's output properties.

channelState :: Maybe ChannelState

Indicates whether the channel is in a running state or not.

fillerSlate :: Maybe SlateSource

Contains information about the slate used to fill gaps between programs in the schedule.

tags :: Maybe (HashMap Text Text)

The tags assigned to the channel.

httpStatus :: Int

The response's http status code.

newUpdateChannelResponse

Arguments

:: Int
$sel:httpStatus:UpdateChannelResponse'

-> UpdateChannelResponse

Create a value of UpdateChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:UpdateChannelResponse', updateChannelResponse_creationTime - The timestamp of when the channel was created.

$sel:arn:UpdateChannelResponse', updateChannelResponse_arn - The ARN of the channel.

$sel:lastModifiedTime:UpdateChannelResponse', updateChannelResponse_lastModifiedTime - The timestamp of when the channel was last modified.

$sel:playbackMode:UpdateChannelResponse', updateChannelResponse_playbackMode - The channel's playback mode.

$sel:channelName:UpdateChannel', updateChannelResponse_channelName - The name of the channel.

$sel:outputs:UpdateChannel', updateChannelResponse_outputs - The channel's output properties.

$sel:channelState:UpdateChannelResponse', updateChannelResponse_channelState - Indicates whether the channel is in a running state or not.

$sel:fillerSlate:UpdateChannelResponse', updateChannelResponse_fillerSlate - Contains information about the slate used to fill gaps between programs in the schedule.

$sel:tags:UpdateChannelResponse', updateChannelResponse_tags - The tags assigned to the channel.

$sel:httpStatus:UpdateChannelResponse', updateChannelResponse_httpStatus - The response's http status code.

Response Lenses

updateChannelResponse_creationTime :: Lens' UpdateChannelResponse (Maybe UTCTime)

The timestamp of when the channel was created.

updateChannelResponse_arn :: Lens' UpdateChannelResponse (Maybe Text)

The ARN of the channel.

updateChannelResponse_lastModifiedTime :: Lens' UpdateChannelResponse (Maybe UTCTime)

The timestamp of when the channel was last modified.

updateChannelResponse_playbackMode :: Lens' UpdateChannelResponse (Maybe Text)

The channel's playback mode.

updateChannelResponse_channelName :: Lens' UpdateChannelResponse (Maybe Text)

The name of the channel.

updateChannelResponse_outputs :: Lens' UpdateChannelResponse (Maybe [ResponseOutputItem])

The channel's output properties.

updateChannelResponse_channelState :: Lens' UpdateChannelResponse (Maybe ChannelState)

Indicates whether the channel is in a running state or not.

updateChannelResponse_fillerSlate :: Lens' UpdateChannelResponse (Maybe SlateSource)

Contains information about the slate used to fill gaps between programs in the schedule.

updateChannelResponse_tags :: Lens' UpdateChannelResponse (Maybe (HashMap Text Text))

The tags assigned to the channel.

updateChannelResponse_httpStatus :: Lens' UpdateChannelResponse Int

The response's http status code.