Network.AWS.MediaTailor.UpdateVodSource

Description

Updates a specific VOD source in a specific source location.

Creating a Request

data UpdateVodSource

See: newUpdateVodSource smart constructor.

Constructors

UpdateVodSource'

Fields

sourceLocationName :: Text

The identifier for the source location you are working on.

vodSourceName :: Text

The identifier for the VOD source you are working on.

httpPackageConfigurations :: [HttpPackageConfiguration]

An array of HTTP package configurations for the VOD source on this account.

newUpdateVodSource

Arguments

:: Text
$sel:sourceLocationName:UpdateVodSource'

-> Text
$sel:vodSourceName:UpdateVodSource'

-> UpdateVodSource

Create a value of UpdateVodSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceLocationName:UpdateVodSource', updateVodSource_sourceLocationName - The identifier for the source location you are working on.

$sel:vodSourceName:UpdateVodSource', updateVodSource_vodSourceName - The identifier for the VOD source you are working on.

$sel:httpPackageConfigurations:UpdateVodSource', updateVodSource_httpPackageConfigurations - An array of HTTP package configurations for the VOD source on this account.

Request Lenses

updateVodSource_sourceLocationName :: Lens' UpdateVodSource Text

The identifier for the source location you are working on.

updateVodSource_vodSourceName :: Lens' UpdateVodSource Text

The identifier for the VOD source you are working on.

updateVodSource_httpPackageConfigurations :: Lens' UpdateVodSource [HttpPackageConfiguration]

An array of HTTP package configurations for the VOD source on this account.

Destructuring the Response

data UpdateVodSourceResponse

See: newUpdateVodSourceResponse smart constructor.

Constructors

UpdateVodSourceResponse'

Fields

creationTime :: Maybe POSIX

The timestamp that indicates when the VOD source was created.

sourceLocationName :: Maybe Text

The name of the source location associated with the VOD source.

arn :: Maybe Text

The ARN of the VOD source.

lastModifiedTime :: Maybe POSIX

The ARN for the VOD source.

httpPackageConfigurations :: Maybe [HttpPackageConfiguration]

The HTTP package configurations.

vodSourceName :: Maybe Text

The name of the VOD source.

tags :: Maybe (HashMap Text Text)

The tags assigned to the VOD source.

httpStatus :: Int

The response's http status code.

newUpdateVodSourceResponse

Arguments

:: Int
$sel:httpStatus:UpdateVodSourceResponse'

-> UpdateVodSourceResponse

Create a value of UpdateVodSourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:UpdateVodSourceResponse', updateVodSourceResponse_creationTime - The timestamp that indicates when the VOD source was created.

$sel:sourceLocationName:UpdateVodSource', updateVodSourceResponse_sourceLocationName - The name of the source location associated with the VOD source.

$sel:arn:UpdateVodSourceResponse', updateVodSourceResponse_arn - The ARN of the VOD source.

$sel:lastModifiedTime:UpdateVodSourceResponse', updateVodSourceResponse_lastModifiedTime - The ARN for the VOD source.

$sel:httpPackageConfigurations:UpdateVodSource', updateVodSourceResponse_httpPackageConfigurations - The HTTP package configurations.

$sel:vodSourceName:UpdateVodSource', updateVodSourceResponse_vodSourceName - The name of the VOD source.

$sel:tags:UpdateVodSourceResponse', updateVodSourceResponse_tags - The tags assigned to the VOD source.

$sel:httpStatus:UpdateVodSourceResponse', updateVodSourceResponse_httpStatus - The response's http status code.

Response Lenses

updateVodSourceResponse_creationTime :: Lens' UpdateVodSourceResponse (Maybe UTCTime)

The timestamp that indicates when the VOD source was created.

updateVodSourceResponse_sourceLocationName :: Lens' UpdateVodSourceResponse (Maybe Text)

The name of the source location associated with the VOD source.

updateVodSourceResponse_arn :: Lens' UpdateVodSourceResponse (Maybe Text)

The ARN of the VOD source.

updateVodSourceResponse_lastModifiedTime :: Lens' UpdateVodSourceResponse (Maybe UTCTime)

The ARN for the VOD source.

updateVodSourceResponse_httpPackageConfigurations :: Lens' UpdateVodSourceResponse (Maybe [HttpPackageConfiguration])

The HTTP package configurations.

updateVodSourceResponse_vodSourceName :: Lens' UpdateVodSourceResponse (Maybe Text)

The name of the VOD source.

updateVodSourceResponse_tags :: Lens' UpdateVodSourceResponse (Maybe (HashMap Text Text))

The tags assigned to the VOD source.

updateVodSourceResponse_httpStatus :: Lens' UpdateVodSourceResponse Int

The response's http status code.