Network.AWS.MemoryDb.BatchUpdateCluster

Description

Apply the service update to a list of clusters supplied. For more information on service updates and applying them, see Applying the service updates.

Creating a Request

data BatchUpdateCluster

See: newBatchUpdateCluster smart constructor.

Constructors

BatchUpdateCluster'

Fields

serviceUpdate :: Maybe ServiceUpdateRequest

The unique ID of the service update

clusterNames :: [Text]

The cluster names to apply the updates.

newBatchUpdateCluster :: BatchUpdateCluster

Create a value of BatchUpdateCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceUpdate:BatchUpdateCluster', batchUpdateCluster_serviceUpdate - The unique ID of the service update

$sel:clusterNames:BatchUpdateCluster', batchUpdateCluster_clusterNames - The cluster names to apply the updates.

Request Lenses

batchUpdateCluster_serviceUpdate :: Lens' BatchUpdateCluster (Maybe ServiceUpdateRequest)

The unique ID of the service update

batchUpdateCluster_clusterNames :: Lens' BatchUpdateCluster [Text]

The cluster names to apply the updates.

Destructuring the Response

data BatchUpdateClusterResponse

See: newBatchUpdateClusterResponse smart constructor.

Constructors

BatchUpdateClusterResponse'

Fields

unprocessedClusters :: Maybe [UnprocessedCluster]

The list of clusters where updates have not been applied.

processedClusters :: Maybe [Cluster]

The list of clusters that have been updated.

httpStatus :: Int

The response's http status code.

newBatchUpdateClusterResponse

Arguments

:: Int
$sel:httpStatus:BatchUpdateClusterResponse'

-> BatchUpdateClusterResponse

Create a value of BatchUpdateClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:unprocessedClusters:BatchUpdateClusterResponse', batchUpdateClusterResponse_unprocessedClusters - The list of clusters where updates have not been applied.

$sel:processedClusters:BatchUpdateClusterResponse', batchUpdateClusterResponse_processedClusters - The list of clusters that have been updated.

$sel:httpStatus:BatchUpdateClusterResponse', batchUpdateClusterResponse_httpStatus - The response's http status code.

Response Lenses

batchUpdateClusterResponse_unprocessedClusters :: Lens' BatchUpdateClusterResponse (Maybe [UnprocessedCluster])

The list of clusters where updates have not been applied.

batchUpdateClusterResponse_processedClusters :: Lens' BatchUpdateClusterResponse (Maybe [Cluster])

The list of clusters that have been updated.

batchUpdateClusterResponse_httpStatus :: Lens' BatchUpdateClusterResponse Int

The response's http status code.