Network.AWS.MemoryDb.CopySnapshot

Description

Makes a copy of an existing snapshot.

Creating a Request

data CopySnapshot

See: newCopySnapshot smart constructor.

Constructors

CopySnapshot'

Fields

targetBucket :: Maybe Text

The Amazon S3 bucket to which the snapshot is exported. This parameter is used only when exporting a snapshot for external access. When using this parameter to export a snapshot, be sure MemoryDB has the needed permissions to this S3 bucket. For more information, see Step 2: Grant MemoryDB Access to Your Amazon S3 Bucket.

kmsKeyId :: Maybe Text

The ID of the KMS key used to encrypt the target snapshot.

tags :: Maybe [Tag]

A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

sourceSnapshotName :: Text

The name of an existing snapshot from which to make a copy.

targetSnapshotName :: Text

A name for the snapshot copy. MemoryDB does not permit overwriting a snapshot, therefore this name must be unique within its context - MemoryDB or an Amazon S3 bucket if exporting.

newCopySnapshot

Arguments

:: Text
$sel:sourceSnapshotName:CopySnapshot'

-> Text
$sel:targetSnapshotName:CopySnapshot'

-> CopySnapshot

Create a value of CopySnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetBucket:CopySnapshot', copySnapshot_targetBucket - The Amazon S3 bucket to which the snapshot is exported. This parameter is used only when exporting a snapshot for external access. When using this parameter to export a snapshot, be sure MemoryDB has the needed permissions to this S3 bucket. For more information, see Step 2: Grant MemoryDB Access to Your Amazon S3 Bucket.

$sel:kmsKeyId:CopySnapshot', copySnapshot_kmsKeyId - The ID of the KMS key used to encrypt the target snapshot.

$sel:tags:CopySnapshot', copySnapshot_tags - A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

$sel:sourceSnapshotName:CopySnapshot', copySnapshot_sourceSnapshotName - The name of an existing snapshot from which to make a copy.

$sel:targetSnapshotName:CopySnapshot', copySnapshot_targetSnapshotName - A name for the snapshot copy. MemoryDB does not permit overwriting a snapshot, therefore this name must be unique within its context - MemoryDB or an Amazon S3 bucket if exporting.

Request Lenses

copySnapshot_targetBucket :: Lens' CopySnapshot (Maybe Text)

The Amazon S3 bucket to which the snapshot is exported. This parameter is used only when exporting a snapshot for external access. When using this parameter to export a snapshot, be sure MemoryDB has the needed permissions to this S3 bucket. For more information, see Step 2: Grant MemoryDB Access to Your Amazon S3 Bucket.

copySnapshot_kmsKeyId :: Lens' CopySnapshot (Maybe Text)

The ID of the KMS key used to encrypt the target snapshot.

copySnapshot_tags :: Lens' CopySnapshot (Maybe [Tag])

A list of tags to be added to this resource. A tag is a key-value pair. A tag key must be accompanied by a tag value, although null is accepted.

copySnapshot_sourceSnapshotName :: Lens' CopySnapshot Text

The name of an existing snapshot from which to make a copy.

copySnapshot_targetSnapshotName :: Lens' CopySnapshot Text

A name for the snapshot copy. MemoryDB does not permit overwriting a snapshot, therefore this name must be unique within its context - MemoryDB or an Amazon S3 bucket if exporting.

Destructuring the Response

data CopySnapshotResponse

See: newCopySnapshotResponse smart constructor.

Constructors

CopySnapshotResponse'

Fields

snapshot :: Maybe Snapshot

Represents a copy of an entire cluster as of the time when the snapshot was taken.

httpStatus :: Int

The response's http status code.

newCopySnapshotResponse

Arguments

:: Int
$sel:httpStatus:CopySnapshotResponse'

-> CopySnapshotResponse

Create a value of CopySnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshot:CopySnapshotResponse', copySnapshotResponse_snapshot - Represents a copy of an entire cluster as of the time when the snapshot was taken.

$sel:httpStatus:CopySnapshotResponse', copySnapshotResponse_httpStatus - The response's http status code.

Response Lenses

copySnapshotResponse_snapshot :: Lens' CopySnapshotResponse (Maybe Snapshot)

Represents a copy of an entire cluster as of the time when the snapshot was taken.

copySnapshotResponse_httpStatus :: Lens' CopySnapshotResponse Int

The response's http status code.