Network.AWS.MemoryDb.DescribeUsers

Description

Returns a list of users.

Creating a Request

data DescribeUsers

See: newDescribeUsers smart constructor.

Constructors

DescribeUsers'

Fields

filters :: Maybe [Filter]

Filter to determine the list of users to return.

userName :: Maybe Text

The name of the user

nextToken :: Maybe Text

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

maxResults :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

newDescribeUsers :: DescribeUsers

Create a value of DescribeUsers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeUsers', describeUsers_filters - Filter to determine the list of users to return.

$sel:userName:DescribeUsers', describeUsers_userName - The name of the user

$sel:nextToken:DescribeUsers', describeUsers_nextToken - An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

$sel:maxResults:DescribeUsers', describeUsers_maxResults - The maximum number of records to include in the response. If more records exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

Request Lenses

describeUsers_filters :: Lens' DescribeUsers (Maybe [Filter])

Filter to determine the list of users to return.

describeUsers_userName :: Lens' DescribeUsers (Maybe Text)

The name of the user

describeUsers_nextToken :: Lens' DescribeUsers (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeUsers_maxResults :: Lens' DescribeUsers (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

Destructuring the Response

data DescribeUsersResponse

See: newDescribeUsersResponse smart constructor.

Constructors

DescribeUsersResponse'

Fields

users :: Maybe [User]

A list of users.

nextToken :: Maybe Text

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

httpStatus :: Int

The response's http status code.

newDescribeUsersResponse

Arguments

:: Int
$sel:httpStatus:DescribeUsersResponse'

-> DescribeUsersResponse

Create a value of DescribeUsersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:users:DescribeUsersResponse', describeUsersResponse_users - A list of users.

$sel:nextToken:DescribeUsers', describeUsersResponse_nextToken - An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

$sel:httpStatus:DescribeUsersResponse', describeUsersResponse_httpStatus - The response's http status code.

Response Lenses

describeUsersResponse_users :: Lens' DescribeUsersResponse (Maybe [User])

A list of users.

describeUsersResponse_nextToken :: Lens' DescribeUsersResponse (Maybe Text)

An optional argument to pass in case the total number of records exceeds the value of MaxResults. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

describeUsersResponse_httpStatus :: Lens' DescribeUsersResponse Int

The response's http status code.