Network.AWS.MemoryDb.Types.ACL

Description

Documentation

data ACL

An Access Control List. You can authenticate users with Access Contol Lists. ACLs enable you to control cluster access by grouping users. These Access control lists are designed as a way to organize access to clusters.

See: newACL smart constructor.

Constructors

ACL'

Fields

status :: Maybe Text

Indicates ACL status. Can be "creating", "active", "modifying", "deleting".

userNames :: Maybe [Text]

The list of user names that belong to the ACL.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the ACL

name :: Maybe Text

The name of the Access Control List

pendingChanges :: Maybe ACLPendingChanges

A list of updates being applied to the ACL.

minimumEngineVersion :: Maybe Text

The minimum engine version supported for the ACL

clusters :: Maybe [Text]

A list of clusters associated with the ACL.

newACL :: ACL

Create a value of ACL with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ACL', acl_status - Indicates ACL status. Can be "creating", "active", "modifying", "deleting".

$sel:userNames:ACL', acl_userNames - The list of user names that belong to the ACL.

$sel:arn:ACL', acl_arn - The Amazon Resource Name (ARN) of the ACL

$sel:name:ACL', acl_name - The name of the Access Control List

$sel:pendingChanges:ACL', acl_pendingChanges - A list of updates being applied to the ACL.

$sel:minimumEngineVersion:ACL', acl_minimumEngineVersion - The minimum engine version supported for the ACL

$sel:clusters:ACL', acl_clusters - A list of clusters associated with the ACL.

acl_status :: Lens' ACL (Maybe Text)

Indicates ACL status. Can be "creating", "active", "modifying", "deleting".

acl_userNames :: Lens' ACL (Maybe [Text])

The list of user names that belong to the ACL.

acl_arn :: Lens' ACL (Maybe Text)

The Amazon Resource Name (ARN) of the ACL

acl_name :: Lens' ACL (Maybe Text)

The name of the Access Control List

acl_pendingChanges :: Lens' ACL (Maybe ACLPendingChanges)

A list of updates being applied to the ACL.

acl_minimumEngineVersion :: Lens' ACL (Maybe Text)

The minimum engine version supported for the ACL

acl_clusters :: Lens' ACL (Maybe [Text])

A list of clusters associated with the ACL.