Network.AWS.MemoryDb.Types.ClusterConfiguration

Description

Documentation

data ClusterConfiguration

A list of cluster configuration options.

See: newClusterConfiguration smart constructor.

Constructors

ClusterConfiguration'

Fields

engineVersion :: Maybe Text

The Redis engine version used by the cluster

vpcId :: Maybe Text

The ID of the VPC the cluster belongs to

snapshotWindow :: Maybe Text

The snapshot window set by the cluster

subnetGroupName :: Maybe Text

The name of the subnet group used by the cluster

numShards :: Maybe Int

The number of shards in the cluster

maintenanceWindow :: Maybe Text

The specified maintenance window for the cluster

topicArn :: Maybe Text

The Amazon Resource Name (ARN) of the SNS notification topic for the cluster

shards :: Maybe [ShardDetail]

The list of shards in the cluster

name :: Maybe Text

The name of the cluster

snapshotRetentionLimit :: Maybe Int

The snapshot retention limit set by the cluster

nodeType :: Maybe Text

The node type used for the cluster

description :: Maybe Text

The description of the cluster configuration

port :: Maybe Int

The port used by the cluster

parameterGroupName :: Maybe Text

The name of parameter group used by the cluster

newClusterConfiguration :: ClusterConfiguration

Create a value of ClusterConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:ClusterConfiguration', clusterConfiguration_engineVersion - The Redis engine version used by the cluster

$sel:vpcId:ClusterConfiguration', clusterConfiguration_vpcId - The ID of the VPC the cluster belongs to

$sel:snapshotWindow:ClusterConfiguration', clusterConfiguration_snapshotWindow - The snapshot window set by the cluster

$sel:subnetGroupName:ClusterConfiguration', clusterConfiguration_subnetGroupName - The name of the subnet group used by the cluster

$sel:numShards:ClusterConfiguration', clusterConfiguration_numShards - The number of shards in the cluster

$sel:maintenanceWindow:ClusterConfiguration', clusterConfiguration_maintenanceWindow - The specified maintenance window for the cluster

$sel:topicArn:ClusterConfiguration', clusterConfiguration_topicArn - The Amazon Resource Name (ARN) of the SNS notification topic for the cluster

$sel:shards:ClusterConfiguration', clusterConfiguration_shards - The list of shards in the cluster

$sel:name:ClusterConfiguration', clusterConfiguration_name - The name of the cluster

$sel:snapshotRetentionLimit:ClusterConfiguration', clusterConfiguration_snapshotRetentionLimit - The snapshot retention limit set by the cluster

$sel:nodeType:ClusterConfiguration', clusterConfiguration_nodeType - The node type used for the cluster

$sel:description:ClusterConfiguration', clusterConfiguration_description - The description of the cluster configuration

$sel:port:ClusterConfiguration', clusterConfiguration_port - The port used by the cluster

$sel:parameterGroupName:ClusterConfiguration', clusterConfiguration_parameterGroupName - The name of parameter group used by the cluster

clusterConfiguration_engineVersion :: Lens' ClusterConfiguration (Maybe Text)

The Redis engine version used by the cluster

clusterConfiguration_vpcId :: Lens' ClusterConfiguration (Maybe Text)

The ID of the VPC the cluster belongs to

clusterConfiguration_snapshotWindow :: Lens' ClusterConfiguration (Maybe Text)

The snapshot window set by the cluster

clusterConfiguration_subnetGroupName :: Lens' ClusterConfiguration (Maybe Text)

The name of the subnet group used by the cluster

clusterConfiguration_numShards :: Lens' ClusterConfiguration (Maybe Int)

The number of shards in the cluster

clusterConfiguration_maintenanceWindow :: Lens' ClusterConfiguration (Maybe Text)

The specified maintenance window for the cluster

clusterConfiguration_topicArn :: Lens' ClusterConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of the SNS notification topic for the cluster

clusterConfiguration_shards :: Lens' ClusterConfiguration (Maybe [ShardDetail])

The list of shards in the cluster

clusterConfiguration_name :: Lens' ClusterConfiguration (Maybe Text)

The name of the cluster

clusterConfiguration_snapshotRetentionLimit :: Lens' ClusterConfiguration (Maybe Int)

The snapshot retention limit set by the cluster

clusterConfiguration_nodeType :: Lens' ClusterConfiguration (Maybe Text)

The node type used for the cluster

clusterConfiguration_description :: Lens' ClusterConfiguration (Maybe Text)

The description of the cluster configuration

clusterConfiguration_port :: Lens' ClusterConfiguration (Maybe Int)

The port used by the cluster

clusterConfiguration_parameterGroupName :: Lens' ClusterConfiguration (Maybe Text)

The name of parameter group used by the cluster