Network.AWS.MemoryDb.Types.ClusterPendingUpdates

Description

Documentation

data ClusterPendingUpdates

A list of updates being applied to the cluster

See: newClusterPendingUpdates smart constructor.

Constructors

ClusterPendingUpdates'

Fields

serviceUpdates :: Maybe [PendingModifiedServiceUpdate]

A list of service updates being applied to the cluster

resharding :: Maybe ReshardingStatus

The status of an online resharding operation.

aCLs :: Maybe ACLsUpdateStatus

A list of ACLs associated with the cluster that are being updated

newClusterPendingUpdates :: ClusterPendingUpdates

Create a value of ClusterPendingUpdates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceUpdates:ClusterPendingUpdates', clusterPendingUpdates_serviceUpdates - A list of service updates being applied to the cluster

$sel:resharding:ClusterPendingUpdates', clusterPendingUpdates_resharding - The status of an online resharding operation.

$sel:aCLs:ClusterPendingUpdates', clusterPendingUpdates_aCLs - A list of ACLs associated with the cluster that are being updated

clusterPendingUpdates_serviceUpdates :: Lens' ClusterPendingUpdates (Maybe [PendingModifiedServiceUpdate])

A list of service updates being applied to the cluster

clusterPendingUpdates_resharding :: Lens' ClusterPendingUpdates (Maybe ReshardingStatus)

The status of an online resharding operation.

clusterPendingUpdates_aCLs :: Lens' ClusterPendingUpdates (Maybe ACLsUpdateStatus)

A list of ACLs associated with the cluster that are being updated