Network.AWS.MemoryDb.Types.ParameterGroup

Description

Documentation

data ParameterGroup

Represents the output of a CreateParameterGroup operation. A parameter group represents a combination of specific values for the parameters that are passed to the engine software during startup.

See: newParameterGroup smart constructor.

Constructors

ParameterGroup'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the parameter group

family :: Maybe Text

The name of the parameter group family that this parameter group is compatible with.

name :: Maybe Text

The name of the parameter group

description :: Maybe Text

A description of the parameter group

newParameterGroup :: ParameterGroup

Create a value of ParameterGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:ParameterGroup', parameterGroup_arn - The Amazon Resource Name (ARN) of the parameter group

$sel:family:ParameterGroup', parameterGroup_family - The name of the parameter group family that this parameter group is compatible with.

$sel:name:ParameterGroup', parameterGroup_name - The name of the parameter group

$sel:description:ParameterGroup', parameterGroup_description - A description of the parameter group

parameterGroup_arn :: Lens' ParameterGroup (Maybe Text)

The Amazon Resource Name (ARN) of the parameter group

parameterGroup_family :: Lens' ParameterGroup (Maybe Text)

The name of the parameter group family that this parameter group is compatible with.

parameterGroup_name :: Lens' ParameterGroup (Maybe Text)

The name of the parameter group

parameterGroup_description :: Lens' ParameterGroup (Maybe Text)

A description of the parameter group