Network.AWS.MemoryDb.Types.SecurityGroupMembership

Description

Documentation

data SecurityGroupMembership

Represents a single security group and its status.

See: newSecurityGroupMembership smart constructor.

Constructors

SecurityGroupMembership'

Fields

status :: Maybe Text

The status of the security group membership. The status changes whenever a security group is modified, or when the security groups assigned to a cluster are modified.

securityGroupId :: Maybe Text

The identifier of the security group.

newSecurityGroupMembership :: SecurityGroupMembership

Create a value of SecurityGroupMembership with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:SecurityGroupMembership', securityGroupMembership_status - The status of the security group membership. The status changes whenever a security group is modified, or when the security groups assigned to a cluster are modified.

$sel:securityGroupId:SecurityGroupMembership', securityGroupMembership_securityGroupId - The identifier of the security group.

securityGroupMembership_status :: Lens' SecurityGroupMembership (Maybe Text)

The status of the security group membership. The status changes whenever a security group is modified, or when the security groups assigned to a cluster are modified.

securityGroupMembership_securityGroupId :: Lens' SecurityGroupMembership (Maybe Text)

The identifier of the security group.