Network.AWS.MemoryDb.Types.Shard

Description

Documentation

data Shard

Represents a collection of nodes in a cluster. One node in the node group is the read/write primary node. All the other nodes are read-only Replica nodes.

See: newShard smart constructor.

Constructors

Shard'

Fields

status :: Maybe Text

The current state of this replication group - creating, available, modifying, deleting.

slots :: Maybe Text

The keyspace for this shard.

numberOfNodes :: Maybe Int

The number of nodes in the shard

name :: Maybe Text

The name of the shard

nodes :: Maybe [Node]

A list containing information about individual nodes within the shard

newShard :: Shard

Create a value of Shard with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Shard', shard_status - The current state of this replication group - creating, available, modifying, deleting.

$sel:slots:Shard', shard_slots - The keyspace for this shard.

$sel:numberOfNodes:Shard', shard_numberOfNodes - The number of nodes in the shard

$sel:name:Shard', shard_name - The name of the shard

$sel:nodes:Shard', shard_nodes - A list containing information about individual nodes within the shard

shard_status :: Lens' Shard (Maybe Text)

The current state of this replication group - creating, available, modifying, deleting.

shard_slots :: Lens' Shard (Maybe Text)

The keyspace for this shard.

shard_numberOfNodes :: Lens' Shard (Maybe Int)

The number of nodes in the shard

shard_name :: Lens' Shard (Maybe Text)

The name of the shard

shard_nodes :: Lens' Shard (Maybe [Node])

A list containing information about individual nodes within the shard