Network.AWS.MemoryDb.Types.SubnetGroup

Description

Documentation

data SubnetGroup

Represents the output of one of the following operations:

CreateSubnetGroup

UpdateSubnetGroup

A subnet group is a collection of subnets (typically private) that you can designate for your clusters running in an Amazon Virtual Private Cloud (VPC) environment.

See: newSubnetGroup smart constructor.

Constructors

SubnetGroup'

Fields

arn :: Maybe Text

The ARN (Amazon Resource Name) of the subnet group.

vpcId :: Maybe Text

The Amazon Virtual Private Cloud identifier (VPC ID) of the subnet group.

subnets :: Maybe [Subnet]

A list of subnets associated with the subnet group.

name :: Maybe Text

The name of the subnet group

description :: Maybe Text

A description of the subnet group

newSubnetGroup :: SubnetGroup

Create a value of SubnetGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:SubnetGroup', subnetGroup_arn - The ARN (Amazon Resource Name) of the subnet group.

$sel:vpcId:SubnetGroup', subnetGroup_vpcId - The Amazon Virtual Private Cloud identifier (VPC ID) of the subnet group.

$sel:subnets:SubnetGroup', subnetGroup_subnets - A list of subnets associated with the subnet group.

$sel:name:SubnetGroup', subnetGroup_name - The name of the subnet group

$sel:description:SubnetGroup', subnetGroup_description - A description of the subnet group

subnetGroup_arn :: Lens' SubnetGroup (Maybe Text)

The ARN (Amazon Resource Name) of the subnet group.

subnetGroup_vpcId :: Lens' SubnetGroup (Maybe Text)

The Amazon Virtual Private Cloud identifier (VPC ID) of the subnet group.

subnetGroup_subnets :: Lens' SubnetGroup (Maybe [Subnet])

A list of subnets associated with the subnet group.

subnetGroup_name :: Lens' SubnetGroup (Maybe Text)

The name of the subnet group

subnetGroup_description :: Lens' SubnetGroup (Maybe Text)

A description of the subnet group