Network.AWS.MemoryDb.Types.User

Description

Documentation

data User

You create users and assign them specific permissions by using an access string. You assign the users to Access Control Lists aligned with a specific role (administrators, human resources) that are then deployed to one or more MemoryDB clusters.

See: newUser smart constructor.

Constructors

User'

Fields

status :: Maybe Text

Indicates the user status. Can be "active", "modifying" or "deleting".

arn :: Maybe Text

The Amazon Resource Name (ARN) of the user.

authentication :: Maybe Authentication

Denotes whether the user requires a password to authenticate.

accessString :: Maybe Text

Access permissions string used for this user.

name :: Maybe Text

The name of the user

aCLNames :: Maybe [Text]

The names of the Access Control Lists to which the user belongs

minimumEngineVersion :: Maybe Text

The minimum engine version supported for the user

newUser :: User

Create a value of User with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:User', user_status - Indicates the user status. Can be "active", "modifying" or "deleting".

$sel:arn:User', user_arn - The Amazon Resource Name (ARN) of the user.

$sel:authentication:User', user_authentication - Denotes whether the user requires a password to authenticate.

$sel:accessString:User', user_accessString - Access permissions string used for this user.

$sel:name:User', user_name - The name of the user

$sel:aCLNames:User', user_aCLNames - The names of the Access Control Lists to which the user belongs

$sel:minimumEngineVersion:User', user_minimumEngineVersion - The minimum engine version supported for the user

user_status :: Lens' User (Maybe Text)

Indicates the user status. Can be "active", "modifying" or "deleting".

user_arn :: Lens' User (Maybe Text)

The Amazon Resource Name (ARN) of the user.

user_authentication :: Lens' User (Maybe Authentication)

Denotes whether the user requires a password to authenticate.

user_accessString :: Lens' User (Maybe Text)

Access permissions string used for this user.

user_name :: Lens' User (Maybe Text)

The name of the user

user_aCLNames :: Lens' User (Maybe [Text])

The names of the Access Control Lists to which the user belongs

user_minimumEngineVersion :: Lens' User (Maybe Text)

The minimum engine version supported for the user