Network.AWS.MemoryDb.UpdateUser

Description

Changes user password(s) and/or access string.

Creating a Request

data UpdateUser

See: newUpdateUser smart constructor.

Constructors

UpdateUser'

Fields

authenticationMode :: Maybe AuthenticationMode

Denotes the user's authentication properties, such as whether it requires a password to authenticate.

accessString :: Maybe Text

Access permissions string used for this user.

userName :: Text

The name of the user

newUpdateUser

Arguments

:: Text
$sel:userName:UpdateUser'

-> UpdateUser

Create a value of UpdateUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:authenticationMode:UpdateUser', updateUser_authenticationMode - Denotes the user's authentication properties, such as whether it requires a password to authenticate.

$sel:accessString:UpdateUser', updateUser_accessString - Access permissions string used for this user.

$sel:userName:UpdateUser', updateUser_userName - The name of the user

Request Lenses

updateUser_authenticationMode :: Lens' UpdateUser (Maybe AuthenticationMode)

Denotes the user's authentication properties, such as whether it requires a password to authenticate.

updateUser_accessString :: Lens' UpdateUser (Maybe Text)

Access permissions string used for this user.

updateUser_userName :: Lens' UpdateUser Text

The name of the user

Destructuring the Response

data UpdateUserResponse

See: newUpdateUserResponse smart constructor.

Constructors

UpdateUserResponse'

Fields

user :: Maybe User

The updated user

httpStatus :: Int

The response's http status code.

newUpdateUserResponse

Arguments

:: Int
$sel:httpStatus:UpdateUserResponse'

-> UpdateUserResponse

Create a value of UpdateUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:user:UpdateUserResponse', updateUserResponse_user - The updated user

$sel:httpStatus:UpdateUserResponse', updateUserResponse_httpStatus - The response's http status code.

Response Lenses

updateUserResponse_user :: Lens' UpdateUserResponse (Maybe User)

The updated user

updateUserResponse_httpStatus :: Lens' UpdateUserResponse Int

The response's http status code.