Network.AWS.MigrationHub

Description

Derived from API version 2017-05-31 of the AWS service descriptions, licensed under Apache 2.0.

The AWS Migration Hub API methods help to obtain server and application migration status and integrate your resource-specific migration tool by providing a programmatic interface to Migration Hub.

Remember that you must set your AWS Migration Hub home region before you call any of these APIs, or a HomeRegionNotSetException error will be returned. Also, you must make the API calls while in your home region.

Service Configuration

defaultService :: Service

API version 2017-05-31 of the Amazon Migration Hub SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by MigrationHub.

AccessDeniedException

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

HomeRegionNotSetException

_HomeRegionNotSetException :: AsError a => Getting (First ServiceError) a ServiceError

The home region is not set. Set the home region to continue.

DryRunOperation

_DryRunOperation :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised to indicate a successfully authorized action when the DryRun flag is set to "true".

PolicyErrorException

_PolicyErrorException :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised when there are problems accessing Application Discovery Service (Application Discovery Service); most likely due to a misconfigured policy or the migrationhub-discovery role is missing or not configured correctly.

ThrottlingException

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied due to request throttling.

InternalServerError

_InternalServerError :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised when an internal, configuration, or dependency error is encountered.

InvalidInputException

_InvalidInputException :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised when the provided input violates a policy constraint or is entered in the wrong format or data type.

ServiceUnavailableException

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised when there is an internal, configuration, or dependency error encountered.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised when the request references a resource (Application Discovery Service configuration, update stream, migration task, etc.) that does not exist in Application Discovery Service (Application Discovery Service) or in Migration Hub's repository.

UnauthorizedOperation

_UnauthorizedOperation :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised to indicate a request was not authorized when the DryRun flag is set to "true".

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

AssociateDiscoveredResource

data AssociateDiscoveredResource

See: newAssociateDiscoveredResource smart constructor.

Constructors

AssociateDiscoveredResource' (Maybe Bool) Text Text DiscoveredResource

newAssociateDiscoveredResource

Arguments

:: Text
$sel:progressUpdateStream:AssociateDiscoveredResource'

-> Text
$sel:migrationTaskName:AssociateDiscoveredResource'

-> DiscoveredResource
$sel:discoveredResource:AssociateDiscoveredResource'

-> AssociateDiscoveredResource

Create a value of AssociateDiscoveredResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:AssociateDiscoveredResource', associateDiscoveredResource_dryRun - Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

$sel:progressUpdateStream:AssociateDiscoveredResource', associateDiscoveredResource_progressUpdateStream - The name of the ProgressUpdateStream.

$sel:migrationTaskName:AssociateDiscoveredResource', associateDiscoveredResource_migrationTaskName - The identifier given to the MigrationTask. /Do not store personal data in this field./

$sel:discoveredResource:AssociateDiscoveredResource', associateDiscoveredResource_discoveredResource - Object representing a Resource.

data AssociateDiscoveredResourceResponse

See: newAssociateDiscoveredResourceResponse smart constructor.

Constructors

AssociateDiscoveredResourceResponse' Int

newAssociateDiscoveredResourceResponse

Arguments

:: Int
$sel:httpStatus:AssociateDiscoveredResourceResponse'

-> AssociateDiscoveredResourceResponse

Create a value of AssociateDiscoveredResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AssociateDiscoveredResourceResponse', associateDiscoveredResourceResponse_httpStatus - The response's http status code.

CreateProgressUpdateStream

data CreateProgressUpdateStream

See: newCreateProgressUpdateStream smart constructor.

Constructors

CreateProgressUpdateStream' (Maybe Bool) Text

newCreateProgressUpdateStream

Arguments

:: Text
$sel:progressUpdateStreamName:CreateProgressUpdateStream'

-> CreateProgressUpdateStream

Create a value of CreateProgressUpdateStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:CreateProgressUpdateStream', createProgressUpdateStream_dryRun - Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

$sel:progressUpdateStreamName:CreateProgressUpdateStream', createProgressUpdateStream_progressUpdateStreamName - The name of the ProgressUpdateStream. /Do not store personal data in this field./

data CreateProgressUpdateStreamResponse

See: newCreateProgressUpdateStreamResponse smart constructor.

Constructors

CreateProgressUpdateStreamResponse' Int

newCreateProgressUpdateStreamResponse

Arguments

:: Int
$sel:httpStatus:CreateProgressUpdateStreamResponse'

-> CreateProgressUpdateStreamResponse

Create a value of CreateProgressUpdateStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateProgressUpdateStreamResponse', createProgressUpdateStreamResponse_httpStatus - The response's http status code.

ListApplicationStates (Paginated)

data ListApplicationStates

See: newListApplicationStates smart constructor.

Constructors

ListApplicationStates' (Maybe Text) (Maybe (NonEmpty Text)) (Maybe Natural)

newListApplicationStates :: ListApplicationStates

Create a value of ListApplicationStates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListApplicationStates', listApplicationStates_nextToken - If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

$sel:applicationIds:ListApplicationStates', listApplicationStates_applicationIds - The configurationIds from the Application Discovery Service that uniquely identifies your applications.

$sel:maxResults:ListApplicationStates', listApplicationStates_maxResults - Maximum number of results to be returned per page.

data ListApplicationStatesResponse

See: newListApplicationStatesResponse smart constructor.

Constructors

ListApplicationStatesResponse' (Maybe [ApplicationState]) (Maybe Text) Int

newListApplicationStatesResponse

Arguments

:: Int
$sel:httpStatus:ListApplicationStatesResponse'

-> ListApplicationStatesResponse

Create a value of ListApplicationStatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationStateList:ListApplicationStatesResponse', listApplicationStatesResponse_applicationStateList - A list of Applications that exist in Application Discovery Service.

$sel:nextToken:ListApplicationStates', listApplicationStatesResponse_nextToken - If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

$sel:httpStatus:ListApplicationStatesResponse', listApplicationStatesResponse_httpStatus - The response's http status code.

ListCreatedArtifacts (Paginated)

data ListCreatedArtifacts

See: newListCreatedArtifacts smart constructor.

Constructors

ListCreatedArtifacts' (Maybe Text) (Maybe Natural) Text Text

newListCreatedArtifacts

Arguments

:: Text
$sel:progressUpdateStream:ListCreatedArtifacts'

-> Text
$sel:migrationTaskName:ListCreatedArtifacts'

-> ListCreatedArtifacts

Create a value of ListCreatedArtifacts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListCreatedArtifacts', listCreatedArtifacts_nextToken - If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

$sel:maxResults:ListCreatedArtifacts', listCreatedArtifacts_maxResults - Maximum number of results to be returned per page.

$sel:progressUpdateStream:ListCreatedArtifacts', listCreatedArtifacts_progressUpdateStream - The name of the ProgressUpdateStream.

$sel:migrationTaskName:ListCreatedArtifacts', listCreatedArtifacts_migrationTaskName - Unique identifier that references the migration task. /Do not store personal data in this field./

data ListCreatedArtifactsResponse

See: newListCreatedArtifactsResponse smart constructor.

Constructors

ListCreatedArtifactsResponse' (Maybe Text) (Maybe [CreatedArtifact]) Int

newListCreatedArtifactsResponse

Arguments

:: Int
$sel:httpStatus:ListCreatedArtifactsResponse'

-> ListCreatedArtifactsResponse

Create a value of ListCreatedArtifactsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListCreatedArtifacts', listCreatedArtifactsResponse_nextToken - If there are more created artifacts than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

$sel:createdArtifactList:ListCreatedArtifactsResponse', listCreatedArtifactsResponse_createdArtifactList - List of created artifacts up to the maximum number of results specified in the request.

$sel:httpStatus:ListCreatedArtifactsResponse', listCreatedArtifactsResponse_httpStatus - The response's http status code.

DisassociateDiscoveredResource

data DisassociateDiscoveredResource

See: newDisassociateDiscoveredResource smart constructor.

Constructors

DisassociateDiscoveredResource' (Maybe Bool) Text Text Text

newDisassociateDiscoveredResource

Arguments

:: Text
$sel:progressUpdateStream:DisassociateDiscoveredResource'

-> Text
$sel:migrationTaskName:DisassociateDiscoveredResource'

-> Text
$sel:configurationId:DisassociateDiscoveredResource'

-> DisassociateDiscoveredResource

Create a value of DisassociateDiscoveredResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DisassociateDiscoveredResource', disassociateDiscoveredResource_dryRun - Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

$sel:progressUpdateStream:DisassociateDiscoveredResource', disassociateDiscoveredResource_progressUpdateStream - The name of the ProgressUpdateStream.

$sel:migrationTaskName:DisassociateDiscoveredResource', disassociateDiscoveredResource_migrationTaskName - The identifier given to the MigrationTask. /Do not store personal data in this field./

$sel:configurationId:DisassociateDiscoveredResource', disassociateDiscoveredResource_configurationId - ConfigurationId of the Application Discovery Service resource to be disassociated.

data DisassociateDiscoveredResourceResponse

See: newDisassociateDiscoveredResourceResponse smart constructor.

Constructors

DisassociateDiscoveredResourceResponse' Int

newDisassociateDiscoveredResourceResponse

Arguments

:: Int
$sel:httpStatus:DisassociateDiscoveredResourceResponse'

-> DisassociateDiscoveredResourceResponse

Create a value of DisassociateDiscoveredResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisassociateDiscoveredResourceResponse', disassociateDiscoveredResourceResponse_httpStatus - The response's http status code.

NotifyApplicationState

data NotifyApplicationState

See: newNotifyApplicationState smart constructor.

Constructors

NotifyApplicationState' (Maybe POSIX) (Maybe Bool) Text ApplicationStatus

newNotifyApplicationState

Arguments

:: Text
$sel:applicationId:NotifyApplicationState'

-> ApplicationStatus
$sel:status:NotifyApplicationState'

-> NotifyApplicationState

Create a value of NotifyApplicationState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateDateTime:NotifyApplicationState', notifyApplicationState_updateDateTime - The timestamp when the application state changed.

$sel:dryRun:NotifyApplicationState', notifyApplicationState_dryRun - Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

$sel:applicationId:NotifyApplicationState', notifyApplicationState_applicationId - The configurationId in Application Discovery Service that uniquely identifies the grouped application.

$sel:status:NotifyApplicationState', notifyApplicationState_status - Status of the application - Not Started, In-Progress, Complete.

data NotifyApplicationStateResponse

See: newNotifyApplicationStateResponse smart constructor.

Constructors

NotifyApplicationStateResponse' Int

newNotifyApplicationStateResponse

Arguments

:: Int
$sel:httpStatus:NotifyApplicationStateResponse'

-> NotifyApplicationStateResponse

Create a value of NotifyApplicationStateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:NotifyApplicationStateResponse', notifyApplicationStateResponse_httpStatus - The response's http status code.

DeleteProgressUpdateStream

data DeleteProgressUpdateStream

See: newDeleteProgressUpdateStream smart constructor.

Constructors

DeleteProgressUpdateStream' (Maybe Bool) Text

newDeleteProgressUpdateStream

Arguments

:: Text
$sel:progressUpdateStreamName:DeleteProgressUpdateStream'

-> DeleteProgressUpdateStream

Create a value of DeleteProgressUpdateStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DeleteProgressUpdateStream', deleteProgressUpdateStream_dryRun - Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

$sel:progressUpdateStreamName:DeleteProgressUpdateStream', deleteProgressUpdateStream_progressUpdateStreamName - The name of the ProgressUpdateStream. /Do not store personal data in this field./

data DeleteProgressUpdateStreamResponse

See: newDeleteProgressUpdateStreamResponse smart constructor.

Constructors

DeleteProgressUpdateStreamResponse' Int

newDeleteProgressUpdateStreamResponse

Arguments

:: Int
$sel:httpStatus:DeleteProgressUpdateStreamResponse'

-> DeleteProgressUpdateStreamResponse

Create a value of DeleteProgressUpdateStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteProgressUpdateStreamResponse', deleteProgressUpdateStreamResponse_httpStatus - The response's http status code.

ListProgressUpdateStreams (Paginated)

data ListProgressUpdateStreams

See: newListProgressUpdateStreams smart constructor.

Constructors

ListProgressUpdateStreams' (Maybe Text) (Maybe Natural)

newListProgressUpdateStreams :: ListProgressUpdateStreams

Create a value of ListProgressUpdateStreams with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListProgressUpdateStreams', listProgressUpdateStreams_nextToken - If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

$sel:maxResults:ListProgressUpdateStreams', listProgressUpdateStreams_maxResults - Filter to limit the maximum number of results to list per page.

data ListProgressUpdateStreamsResponse

See: newListProgressUpdateStreamsResponse smart constructor.

Constructors

ListProgressUpdateStreamsResponse' (Maybe [ProgressUpdateStreamSummary]) (Maybe Text) Int

newListProgressUpdateStreamsResponse

Arguments

:: Int
$sel:httpStatus:ListProgressUpdateStreamsResponse'

-> ListProgressUpdateStreamsResponse

Create a value of ListProgressUpdateStreamsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:progressUpdateStreamSummaryList:ListProgressUpdateStreamsResponse', listProgressUpdateStreamsResponse_progressUpdateStreamSummaryList - List of progress update streams up to the max number of results passed in the input.

$sel:nextToken:ListProgressUpdateStreams', listProgressUpdateStreamsResponse_nextToken - If there are more streams created than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

$sel:httpStatus:ListProgressUpdateStreamsResponse', listProgressUpdateStreamsResponse_httpStatus - The response's http status code.

DisassociateCreatedArtifact

data DisassociateCreatedArtifact

See: newDisassociateCreatedArtifact smart constructor.

Constructors

DisassociateCreatedArtifact' (Maybe Bool) Text Text Text

newDisassociateCreatedArtifact

Arguments

:: Text
$sel:progressUpdateStream:DisassociateCreatedArtifact'

-> Text
$sel:migrationTaskName:DisassociateCreatedArtifact'

-> Text
$sel:createdArtifactName:DisassociateCreatedArtifact'

-> DisassociateCreatedArtifact

Create a value of DisassociateCreatedArtifact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DisassociateCreatedArtifact', disassociateCreatedArtifact_dryRun - Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

$sel:progressUpdateStream:DisassociateCreatedArtifact', disassociateCreatedArtifact_progressUpdateStream - The name of the ProgressUpdateStream.

$sel:migrationTaskName:DisassociateCreatedArtifact', disassociateCreatedArtifact_migrationTaskName - Unique identifier that references the migration task to be disassociated with the artifact. Do not store personal data in this field.

$sel:createdArtifactName:DisassociateCreatedArtifact', disassociateCreatedArtifact_createdArtifactName - An ARN of the AWS resource related to the migration (e.g., AMI, EC2 instance, RDS instance, etc.)

data DisassociateCreatedArtifactResponse

See: newDisassociateCreatedArtifactResponse smart constructor.

Constructors

DisassociateCreatedArtifactResponse' Int

newDisassociateCreatedArtifactResponse

Arguments

:: Int
$sel:httpStatus:DisassociateCreatedArtifactResponse'

-> DisassociateCreatedArtifactResponse

Create a value of DisassociateCreatedArtifactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DisassociateCreatedArtifactResponse', disassociateCreatedArtifactResponse_httpStatus - The response's http status code.

ImportMigrationTask

data ImportMigrationTask

See: newImportMigrationTask smart constructor.

Constructors

ImportMigrationTask' (Maybe Bool) Text Text

newImportMigrationTask

Arguments

:: Text
$sel:progressUpdateStream:ImportMigrationTask'

-> Text
$sel:migrationTaskName:ImportMigrationTask'

-> ImportMigrationTask

Create a value of ImportMigrationTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:ImportMigrationTask', importMigrationTask_dryRun - Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

$sel:progressUpdateStream:ImportMigrationTask', importMigrationTask_progressUpdateStream - The name of the ProgressUpdateStream. >

$sel:migrationTaskName:ImportMigrationTask', importMigrationTask_migrationTaskName - Unique identifier that references the migration task. /Do not store personal data in this field./

data ImportMigrationTaskResponse

See: newImportMigrationTaskResponse smart constructor.

Constructors

ImportMigrationTaskResponse' Int

newImportMigrationTaskResponse

Arguments

:: Int
$sel:httpStatus:ImportMigrationTaskResponse'

-> ImportMigrationTaskResponse

Create a value of ImportMigrationTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ImportMigrationTaskResponse', importMigrationTaskResponse_httpStatus - The response's http status code.

DescribeMigrationTask

data DescribeMigrationTask

See: newDescribeMigrationTask smart constructor.

Constructors

DescribeMigrationTask' Text Text

newDescribeMigrationTask

Arguments

:: Text
$sel:progressUpdateStream:DescribeMigrationTask'

-> Text
$sel:migrationTaskName:DescribeMigrationTask'

-> DescribeMigrationTask

Create a value of DescribeMigrationTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:progressUpdateStream:DescribeMigrationTask', describeMigrationTask_progressUpdateStream - The name of the ProgressUpdateStream.

$sel:migrationTaskName:DescribeMigrationTask', describeMigrationTask_migrationTaskName - The identifier given to the MigrationTask. /Do not store personal data in this field./

data DescribeMigrationTaskResponse

See: newDescribeMigrationTaskResponse smart constructor.

Constructors

DescribeMigrationTaskResponse' (Maybe MigrationTask) Int

newDescribeMigrationTaskResponse

Arguments

:: Int
$sel:httpStatus:DescribeMigrationTaskResponse'

-> DescribeMigrationTaskResponse

Create a value of DescribeMigrationTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:migrationTask:DescribeMigrationTaskResponse', describeMigrationTaskResponse_migrationTask - Object encapsulating information about the migration task.

$sel:httpStatus:DescribeMigrationTaskResponse', describeMigrationTaskResponse_httpStatus - The response's http status code.

PutResourceAttributes

data PutResourceAttributes

See: newPutResourceAttributes smart constructor.

Constructors

PutResourceAttributes' (Maybe Bool) Text Text (NonEmpty ResourceAttribute)

newPutResourceAttributes

Arguments

:: Text
$sel:progressUpdateStream:PutResourceAttributes'

-> Text
$sel:migrationTaskName:PutResourceAttributes'

-> NonEmpty ResourceAttribute
$sel:resourceAttributeList:PutResourceAttributes'

-> PutResourceAttributes

Create a value of PutResourceAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:PutResourceAttributes', putResourceAttributes_dryRun - Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

$sel:progressUpdateStream:PutResourceAttributes', putResourceAttributes_progressUpdateStream - The name of the ProgressUpdateStream.

$sel:migrationTaskName:PutResourceAttributes', putResourceAttributes_migrationTaskName - Unique identifier that references the migration task. /Do not store personal data in this field./

$sel:resourceAttributeList:PutResourceAttributes', putResourceAttributes_resourceAttributeList - Information about the resource that is being migrated. This data will be used to map the task to a resource in the Application Discovery Service repository.

Takes the object array of ResourceAttribute where the Type field is reserved for the following values: IPV4_ADDRESS | IPV6_ADDRESS | MAC_ADDRESS | FQDN | VM_MANAGER_ID | VM_MANAGED_OBJECT_REFERENCE | VM_NAME | VM_PATH | BIOS_ID | MOTHERBOARD_SERIAL_NUMBER where the identifying value can be a string up to 256 characters.

If any "VM" related value is set for a ResourceAttribute object, it is required that VM_MANAGER_ID, as a minimum, is always set. If VM_MANAGER_ID is not set, then all "VM" fields will be discarded and "VM" fields will not be used for matching the migration task to a server in Application Discovery Service repository. See the Example section below for a use case of specifying "VM" related values.

If a server you are trying to match has multiple IP or MAC addresses, you should provide as many as you know in separate type/value pairs passed to the ResourceAttributeList parameter to maximize the chances of matching.

data PutResourceAttributesResponse

See: newPutResourceAttributesResponse smart constructor.

Constructors

PutResourceAttributesResponse' Int

newPutResourceAttributesResponse

Arguments

:: Int
$sel:httpStatus:PutResourceAttributesResponse'

-> PutResourceAttributesResponse

Create a value of PutResourceAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutResourceAttributesResponse', putResourceAttributesResponse_httpStatus - The response's http status code.

AssociateCreatedArtifact

data AssociateCreatedArtifact

See: newAssociateCreatedArtifact smart constructor.

Constructors

AssociateCreatedArtifact' (Maybe Bool) Text Text CreatedArtifact

newAssociateCreatedArtifact

Arguments

:: Text
$sel:progressUpdateStream:AssociateCreatedArtifact'

-> Text
$sel:migrationTaskName:AssociateCreatedArtifact'

-> CreatedArtifact
$sel:createdArtifact:AssociateCreatedArtifact'

-> AssociateCreatedArtifact

Create a value of AssociateCreatedArtifact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:AssociateCreatedArtifact', associateCreatedArtifact_dryRun - Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

$sel:progressUpdateStream:AssociateCreatedArtifact', associateCreatedArtifact_progressUpdateStream - The name of the ProgressUpdateStream.

$sel:migrationTaskName:AssociateCreatedArtifact', associateCreatedArtifact_migrationTaskName - Unique identifier that references the migration task. /Do not store personal data in this field./

$sel:createdArtifact:AssociateCreatedArtifact', associateCreatedArtifact_createdArtifact - An ARN of the AWS resource related to the migration (e.g., AMI, EC2 instance, RDS instance, etc.)

data AssociateCreatedArtifactResponse

See: newAssociateCreatedArtifactResponse smart constructor.

Constructors

AssociateCreatedArtifactResponse' Int

newAssociateCreatedArtifactResponse

Arguments

:: Int
$sel:httpStatus:AssociateCreatedArtifactResponse'

-> AssociateCreatedArtifactResponse

Create a value of AssociateCreatedArtifactResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:AssociateCreatedArtifactResponse', associateCreatedArtifactResponse_httpStatus - The response's http status code.

NotifyMigrationTaskState

data NotifyMigrationTaskState

See: newNotifyMigrationTaskState smart constructor.

Constructors

NotifyMigrationTaskState' (Maybe Bool) Text Text Task POSIX Natural

newNotifyMigrationTaskState

Arguments

:: Text
$sel:progressUpdateStream:NotifyMigrationTaskState'

-> Text
$sel:migrationTaskName:NotifyMigrationTaskState'

-> Task
$sel:task:NotifyMigrationTaskState'

-> UTCTime
$sel:updateDateTime:NotifyMigrationTaskState'

-> Natural
$sel:nextUpdateSeconds:NotifyMigrationTaskState'

-> NotifyMigrationTaskState

Create a value of NotifyMigrationTaskState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:NotifyMigrationTaskState', notifyMigrationTaskState_dryRun - Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

$sel:progressUpdateStream:NotifyMigrationTaskState', notifyMigrationTaskState_progressUpdateStream - The name of the ProgressUpdateStream.

$sel:migrationTaskName:NotifyMigrationTaskState', notifyMigrationTaskState_migrationTaskName - Unique identifier that references the migration task. /Do not store personal data in this field./

$sel:task:NotifyMigrationTaskState', notifyMigrationTaskState_task - Information about the task's progress and status.

$sel:updateDateTime:NotifyMigrationTaskState', notifyMigrationTaskState_updateDateTime - The timestamp when the task was gathered.

$sel:nextUpdateSeconds:NotifyMigrationTaskState', notifyMigrationTaskState_nextUpdateSeconds - Number of seconds after the UpdateDateTime within which the Migration Hub can expect an update. If Migration Hub does not receive an update within the specified interval, then the migration task will be considered stale.

data NotifyMigrationTaskStateResponse

See: newNotifyMigrationTaskStateResponse smart constructor.

Constructors

NotifyMigrationTaskStateResponse' Int

newNotifyMigrationTaskStateResponse

Arguments

:: Int
$sel:httpStatus:NotifyMigrationTaskStateResponse'

-> NotifyMigrationTaskStateResponse

Create a value of NotifyMigrationTaskStateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:NotifyMigrationTaskStateResponse', notifyMigrationTaskStateResponse_httpStatus - The response's http status code.

DescribeApplicationState

data DescribeApplicationState

See: newDescribeApplicationState smart constructor.

Constructors

DescribeApplicationState' Text

newDescribeApplicationState

Arguments

:: Text
$sel:applicationId:DescribeApplicationState'

-> DescribeApplicationState

Create a value of DescribeApplicationState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:DescribeApplicationState', describeApplicationState_applicationId - The configurationId in Application Discovery Service that uniquely identifies the grouped application.

data DescribeApplicationStateResponse

See: newDescribeApplicationStateResponse smart constructor.

Constructors

DescribeApplicationStateResponse' (Maybe POSIX) (Maybe ApplicationStatus) Int

newDescribeApplicationStateResponse

Arguments

:: Int
$sel:httpStatus:DescribeApplicationStateResponse'

-> DescribeApplicationStateResponse

Create a value of DescribeApplicationStateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedTime:DescribeApplicationStateResponse', describeApplicationStateResponse_lastUpdatedTime - The timestamp when the application status was last updated.

$sel:applicationStatus:DescribeApplicationStateResponse', describeApplicationStateResponse_applicationStatus - Status of the application - Not Started, In-Progress, Complete.

$sel:httpStatus:DescribeApplicationStateResponse', describeApplicationStateResponse_httpStatus - The response's http status code.

ListMigrationTasks (Paginated)

data ListMigrationTasks

See: newListMigrationTasks smart constructor.

Constructors

ListMigrationTasks' (Maybe Text) (Maybe Text) (Maybe Natural)

newListMigrationTasks :: ListMigrationTasks

Create a value of ListMigrationTasks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceName:ListMigrationTasks', listMigrationTasks_resourceName - Filter migration tasks by discovered resource name.

$sel:nextToken:ListMigrationTasks', listMigrationTasks_nextToken - If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

$sel:maxResults:ListMigrationTasks', listMigrationTasks_maxResults - Value to specify how many results are returned per page.

data ListMigrationTasksResponse

See: newListMigrationTasksResponse smart constructor.

Constructors

ListMigrationTasksResponse' (Maybe [MigrationTaskSummary]) (Maybe Text) Int

newListMigrationTasksResponse

Arguments

:: Int
$sel:httpStatus:ListMigrationTasksResponse'

-> ListMigrationTasksResponse

Create a value of ListMigrationTasksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:migrationTaskSummaryList:ListMigrationTasksResponse', listMigrationTasksResponse_migrationTaskSummaryList - Lists the migration task's summary which includes: MigrationTaskName, ProgressPercent, ProgressUpdateStream, Status, and the UpdateDateTime for each task.

$sel:nextToken:ListMigrationTasks', listMigrationTasksResponse_nextToken - If there are more migration tasks than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

$sel:httpStatus:ListMigrationTasksResponse', listMigrationTasksResponse_httpStatus - The response's http status code.

ListDiscoveredResources (Paginated)

data ListDiscoveredResources

See: newListDiscoveredResources smart constructor.

Constructors

ListDiscoveredResources' (Maybe Text) (Maybe Natural) Text Text

newListDiscoveredResources

Arguments

:: Text
$sel:progressUpdateStream:ListDiscoveredResources'

-> Text
$sel:migrationTaskName:ListDiscoveredResources'

-> ListDiscoveredResources

Create a value of ListDiscoveredResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDiscoveredResources', listDiscoveredResources_nextToken - If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

$sel:maxResults:ListDiscoveredResources', listDiscoveredResources_maxResults - The maximum number of results returned per page.

$sel:progressUpdateStream:ListDiscoveredResources', listDiscoveredResources_progressUpdateStream - The name of the ProgressUpdateStream.

$sel:migrationTaskName:ListDiscoveredResources', listDiscoveredResources_migrationTaskName - The name of the MigrationTask. /Do not store personal data in this field./

data ListDiscoveredResourcesResponse

See: newListDiscoveredResourcesResponse smart constructor.

Constructors

ListDiscoveredResourcesResponse' (Maybe [DiscoveredResource]) (Maybe Text) Int

newListDiscoveredResourcesResponse

Arguments

:: Int
$sel:httpStatus:ListDiscoveredResourcesResponse'

-> ListDiscoveredResourcesResponse

Create a value of ListDiscoveredResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:discoveredResourceList:ListDiscoveredResourcesResponse', listDiscoveredResourcesResponse_discoveredResourceList - Returned list of discovered resources associated with the given MigrationTask.

$sel:nextToken:ListDiscoveredResources', listDiscoveredResourcesResponse_nextToken - If there are more discovered resources than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

$sel:httpStatus:ListDiscoveredResourcesResponse', listDiscoveredResourcesResponse_httpStatus - The response's http status code.

Types

ApplicationStatus

newtype ApplicationStatus

Constructors

ApplicationStatus'

Fields

fromApplicationStatus :: Text

Bundled Patterns

pattern ApplicationStatus_COMPLETED :: ApplicationStatus

pattern ApplicationStatus_IN_PROGRESS :: ApplicationStatus

pattern ApplicationStatus_NOT_STARTED :: ApplicationStatus

MigrationStatus

newtype MigrationStatus

Constructors

MigrationStatus'

Fields

fromMigrationStatus :: Text

Bundled Patterns

pattern MigrationStatus_COMPLETED :: MigrationStatus

pattern MigrationStatus_FAILED :: MigrationStatus

pattern MigrationStatus_IN_PROGRESS :: MigrationStatus

pattern MigrationStatus_NOT_STARTED :: MigrationStatus

ResourceAttributeType

newtype ResourceAttributeType

Constructors

ResourceAttributeType'

Fields

fromResourceAttributeType :: Text

Bundled Patterns

pattern ResourceAttributeType_BIOS_ID :: ResourceAttributeType

pattern ResourceAttributeType_FQDN :: ResourceAttributeType

pattern ResourceAttributeType_IPV4_ADDRESS :: ResourceAttributeType

pattern ResourceAttributeType_IPV6_ADDRESS :: ResourceAttributeType

pattern ResourceAttributeType_MAC_ADDRESS :: ResourceAttributeType

pattern ResourceAttributeType_MOTHERBOARD_SERIAL_NUMBER :: ResourceAttributeType

pattern ResourceAttributeType_VM_MANAGED_OBJECT_REFERENCE :: ResourceAttributeType

pattern ResourceAttributeType_VM_MANAGER_ID :: ResourceAttributeType

pattern ResourceAttributeType_VM_NAME :: ResourceAttributeType

pattern ResourceAttributeType_VM_PATH :: ResourceAttributeType

ApplicationState

data ApplicationState

The state of an application discovered through Migration Hub import, the AWS Agentless Discovery Connector, or the AWS Application Discovery Agent.

See: newApplicationState smart constructor.

Constructors

ApplicationState' (Maybe POSIX) (Maybe Text) (Maybe ApplicationStatus)

newApplicationState :: ApplicationState

Create a value of ApplicationState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedTime:ApplicationState', applicationState_lastUpdatedTime - The timestamp when the application status was last updated.

$sel:applicationId:ApplicationState', applicationState_applicationId - The configurationId from the Application Discovery Service that uniquely identifies an application.

$sel:applicationStatus:ApplicationState', applicationState_applicationStatus - The current status of an application.

CreatedArtifact

data CreatedArtifact

An ARN of the AWS cloud resource target receiving the migration (e.g., AMI, EC2 instance, RDS instance, etc.).

See: newCreatedArtifact smart constructor.

Constructors

CreatedArtifact' (Maybe Text) Text

newCreatedArtifact

Arguments

:: Text
$sel:name:CreatedArtifact'

-> CreatedArtifact

Create a value of CreatedArtifact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreatedArtifact', createdArtifact_description - A description that can be free-form text to record additional detail about the artifact for clarity or for later reference.

$sel:name:CreatedArtifact', createdArtifact_name - An ARN that uniquely identifies the result of a migration task.

DiscoveredResource

data DiscoveredResource

Object representing the on-premises resource being migrated.

See: newDiscoveredResource smart constructor.

Constructors

DiscoveredResource' (Maybe Text) Text

newDiscoveredResource

Arguments

:: Text
$sel:configurationId:DiscoveredResource'

-> DiscoveredResource

Create a value of DiscoveredResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:DiscoveredResource', discoveredResource_description - A description that can be free-form text to record additional detail about the discovered resource for clarity or later reference.

$sel:configurationId:DiscoveredResource', discoveredResource_configurationId - The configurationId in Application Discovery Service that uniquely identifies the on-premise resource.

MigrationTask

data MigrationTask

Represents a migration task in a migration tool.

See: newMigrationTask smart constructor.

Constructors

MigrationTask' (Maybe POSIX) (Maybe [ResourceAttribute]) (Maybe Task) (Maybe Text) (Maybe Text)

newMigrationTask :: MigrationTask

Create a value of MigrationTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateDateTime:MigrationTask', migrationTask_updateDateTime - The timestamp when the task was gathered.

$sel:resourceAttributeList:MigrationTask', migrationTask_resourceAttributeList - Information about the resource that is being migrated. This data will be used to map the task to a resource in the Application Discovery Service repository.

$sel:task:MigrationTask', migrationTask_task - Task object encapsulating task information.

$sel:progressUpdateStream:MigrationTask', migrationTask_progressUpdateStream - A name that identifies the vendor of the migration tool being used.

$sel:migrationTaskName:MigrationTask', migrationTask_migrationTaskName - Unique identifier that references the migration task. /Do not store personal data in this field./

MigrationTaskSummary

data MigrationTaskSummary

MigrationTaskSummary includes MigrationTaskName, ProgressPercent, ProgressUpdateStream, Status, and UpdateDateTime for each task.

See: newMigrationTaskSummary smart constructor.

Constructors

MigrationTaskSummary' (Maybe MigrationStatus) (Maybe POSIX) (Maybe Natural) (Maybe Text) (Maybe Text) (Maybe Text)

newMigrationTaskSummary :: MigrationTaskSummary

Create a value of MigrationTaskSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:MigrationTaskSummary', migrationTaskSummary_status - Status of the task.

$sel:updateDateTime:MigrationTaskSummary', migrationTaskSummary_updateDateTime - The timestamp when the task was gathered.

$sel:progressPercent:MigrationTaskSummary', migrationTaskSummary_progressPercent - Indication of the percentage completion of the task.

$sel:statusDetail:MigrationTaskSummary', migrationTaskSummary_statusDetail - Detail information of what is being done within the overall status state.

$sel:progressUpdateStream:MigrationTaskSummary', migrationTaskSummary_progressUpdateStream - An AWS resource used for access control. It should uniquely identify the migration tool as it is used for all updates made by the tool.

$sel:migrationTaskName:MigrationTaskSummary', migrationTaskSummary_migrationTaskName - Unique identifier that references the migration task. /Do not store personal data in this field./

ProgressUpdateStreamSummary

data ProgressUpdateStreamSummary

Summary of the AWS resource used for access control that is implicitly linked to your AWS account.

See: newProgressUpdateStreamSummary smart constructor.

Constructors

ProgressUpdateStreamSummary' (Maybe Text)

newProgressUpdateStreamSummary :: ProgressUpdateStreamSummary

Create a value of ProgressUpdateStreamSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:progressUpdateStreamName:ProgressUpdateStreamSummary', progressUpdateStreamSummary_progressUpdateStreamName - The name of the ProgressUpdateStream. /Do not store personal data in this field./

ResourceAttribute

data ResourceAttribute

Attribute associated with a resource.

Note the corresponding format required per type listed below:

IPV4

x.x.x.x
where x is an integer in the range [0,255]

IPV6

y : y : y : y : y : y : y : y
where y is a hexadecimal between 0 and FFFF. [0, FFFF]

MAC_ADDRESS

^([0-9A-Fa-f]{2}[:-]){5}([0-9A-Fa-f]{2})$

FQDN

^[^<>{}\\\\/?,=\\p{Cntrl}]{1,256}$

See: newResourceAttribute smart constructor.

Constructors

ResourceAttribute' ResourceAttributeType Text

newResourceAttribute

Arguments

:: ResourceAttributeType
$sel:type':ResourceAttribute'

-> Text
$sel:value:ResourceAttribute'

-> ResourceAttribute

Create a value of ResourceAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':ResourceAttribute', resourceAttribute_type - Type of resource.

$sel:value:ResourceAttribute', resourceAttribute_value - Value of the resource type.

Task

data Task

Task object encapsulating task information.

See: newTask smart constructor.

Constructors

Task' (Maybe Natural) (Maybe Text) MigrationStatus

newTask

Arguments

:: MigrationStatus
$sel:status:Task'

-> Task

Create a value of Task with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:progressPercent:Task', task_progressPercent - Indication of the percentage completion of the task.

$sel:statusDetail:Task', task_statusDetail - Details of task status as notified by a migration tool. A tool might use this field to provide clarifying information about the status that is unique to that tool or that explains an error state.

$sel:status:Task', task_status - Status of the task - Not Started, In-Progress, Complete.