Network.AWS.MigrationHub.CreateProgressUpdateStream

Description

Creates a progress update stream which is an AWS resource used for access control as well as a namespace for migration task names that is implicitly linked to your AWS account. It must uniquely identify the migration tool as it is used for all updates made by the tool; however, it does not need to be unique for each AWS account because it is scoped to the AWS account.

Creating a Request

data CreateProgressUpdateStream

See: newCreateProgressUpdateStream smart constructor.

Constructors

CreateProgressUpdateStream'

Fields

dryRun :: Maybe Bool

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

progressUpdateStreamName :: Text

The name of the ProgressUpdateStream. /Do not store personal data in this field./

newCreateProgressUpdateStream

Arguments

:: Text
$sel:progressUpdateStreamName:CreateProgressUpdateStream'

-> CreateProgressUpdateStream

Create a value of CreateProgressUpdateStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:CreateProgressUpdateStream', createProgressUpdateStream_dryRun - Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

$sel:progressUpdateStreamName:CreateProgressUpdateStream', createProgressUpdateStream_progressUpdateStreamName - The name of the ProgressUpdateStream. /Do not store personal data in this field./

Request Lenses

createProgressUpdateStream_dryRun :: Lens' CreateProgressUpdateStream (Maybe Bool)

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

createProgressUpdateStream_progressUpdateStreamName :: Lens' CreateProgressUpdateStream Text

The name of the ProgressUpdateStream. /Do not store personal data in this field./

Destructuring the Response

data CreateProgressUpdateStreamResponse

See: newCreateProgressUpdateStreamResponse smart constructor.

Constructors

CreateProgressUpdateStreamResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateProgressUpdateStreamResponse

Arguments

:: Int
$sel:httpStatus:CreateProgressUpdateStreamResponse'

-> CreateProgressUpdateStreamResponse

Create a value of CreateProgressUpdateStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateProgressUpdateStreamResponse', createProgressUpdateStreamResponse_httpStatus - The response's http status code.

Response Lenses

createProgressUpdateStreamResponse_httpStatus :: Lens' CreateProgressUpdateStreamResponse Int

The response's http status code.