Network.AWS.MigrationHub.DeleteProgressUpdateStream

Description

Deletes a progress update stream, including all of its tasks, which was previously created as an AWS resource used for access control. This API has the following traits:

The only parameter needed for DeleteProgressUpdateStream is the stream name (same as a CreateProgressUpdateStream call).

The call will return, and a background process will asynchronously delete the stream and all of its resources (tasks, associated resources, resource attributes, created artifacts).

If the stream takes time to be deleted, it might still show up on a ListProgressUpdateStreams call.

CreateProgressUpdateStream, ImportMigrationTask, NotifyMigrationTaskState, and all Associate[*] APIs related to the tasks belonging to the stream will throw "InvalidInputException" if the stream of the same name is in the process of being deleted.

Once the stream and all of its resources are deleted, CreateProgressUpdateStream for a stream of the same name will succeed, and that stream will be an entirely new logical resource (without any resources associated with the old stream).

Creating a Request

data DeleteProgressUpdateStream

See: newDeleteProgressUpdateStream smart constructor.

Constructors

DeleteProgressUpdateStream'

Fields

dryRun :: Maybe Bool

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

progressUpdateStreamName :: Text

The name of the ProgressUpdateStream. /Do not store personal data in this field./

newDeleteProgressUpdateStream

Arguments

:: Text
$sel:progressUpdateStreamName:DeleteProgressUpdateStream'

-> DeleteProgressUpdateStream

Create a value of DeleteProgressUpdateStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:DeleteProgressUpdateStream', deleteProgressUpdateStream_dryRun - Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

$sel:progressUpdateStreamName:DeleteProgressUpdateStream', deleteProgressUpdateStream_progressUpdateStreamName - The name of the ProgressUpdateStream. /Do not store personal data in this field./

Request Lenses

deleteProgressUpdateStream_dryRun :: Lens' DeleteProgressUpdateStream (Maybe Bool)

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

deleteProgressUpdateStream_progressUpdateStreamName :: Lens' DeleteProgressUpdateStream Text

The name of the ProgressUpdateStream. /Do not store personal data in this field./

Destructuring the Response

data DeleteProgressUpdateStreamResponse

See: newDeleteProgressUpdateStreamResponse smart constructor.

Constructors

DeleteProgressUpdateStreamResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteProgressUpdateStreamResponse

Arguments

:: Int
$sel:httpStatus:DeleteProgressUpdateStreamResponse'

-> DeleteProgressUpdateStreamResponse

Create a value of DeleteProgressUpdateStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteProgressUpdateStreamResponse', deleteProgressUpdateStreamResponse_httpStatus - The response's http status code.

Response Lenses

deleteProgressUpdateStreamResponse_httpStatus :: Lens' DeleteProgressUpdateStreamResponse Int

The response's http status code.