Network.AWS.MigrationHub.ListCreatedArtifacts

Description

Lists the created artifacts attached to a given migration task in an update stream. This API has the following traits:

Gets the list of the created artifacts while migration is taking place.

Shows the artifacts created by the migration tool that was associated by the AssociateCreatedArtifact API.

Lists created artifacts in a paginated interface.

This operation returns paginated results.

Creating a Request

data ListCreatedArtifacts

See: newListCreatedArtifacts smart constructor.

Constructors

ListCreatedArtifacts'

Fields

nextToken :: Maybe Text

If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

maxResults :: Maybe Natural

Maximum number of results to be returned per page.

progressUpdateStream :: Text

The name of the ProgressUpdateStream.

migrationTaskName :: Text

Unique identifier that references the migration task. /Do not store personal data in this field./

newListCreatedArtifacts

Arguments

:: Text
$sel:progressUpdateStream:ListCreatedArtifacts'

-> Text
$sel:migrationTaskName:ListCreatedArtifacts'

-> ListCreatedArtifacts

Create a value of ListCreatedArtifacts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListCreatedArtifacts', listCreatedArtifacts_nextToken - If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

$sel:maxResults:ListCreatedArtifacts', listCreatedArtifacts_maxResults - Maximum number of results to be returned per page.

$sel:progressUpdateStream:ListCreatedArtifacts', listCreatedArtifacts_progressUpdateStream - The name of the ProgressUpdateStream.

$sel:migrationTaskName:ListCreatedArtifacts', listCreatedArtifacts_migrationTaskName - Unique identifier that references the migration task. /Do not store personal data in this field./

Request Lenses

listCreatedArtifacts_nextToken :: Lens' ListCreatedArtifacts (Maybe Text)

If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

listCreatedArtifacts_maxResults :: Lens' ListCreatedArtifacts (Maybe Natural)

Maximum number of results to be returned per page.

listCreatedArtifacts_progressUpdateStream :: Lens' ListCreatedArtifacts Text

The name of the ProgressUpdateStream.

listCreatedArtifacts_migrationTaskName :: Lens' ListCreatedArtifacts Text

Unique identifier that references the migration task. /Do not store personal data in this field./

Destructuring the Response

data ListCreatedArtifactsResponse

See: newListCreatedArtifactsResponse smart constructor.

Constructors

ListCreatedArtifactsResponse'

Fields

nextToken :: Maybe Text

If there are more created artifacts than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

createdArtifactList :: Maybe [CreatedArtifact]

List of created artifacts up to the maximum number of results specified in the request.

httpStatus :: Int

The response's http status code.

newListCreatedArtifactsResponse

Arguments

:: Int
$sel:httpStatus:ListCreatedArtifactsResponse'

-> ListCreatedArtifactsResponse

Create a value of ListCreatedArtifactsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListCreatedArtifacts', listCreatedArtifactsResponse_nextToken - If there are more created artifacts than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

$sel:createdArtifactList:ListCreatedArtifactsResponse', listCreatedArtifactsResponse_createdArtifactList - List of created artifacts up to the maximum number of results specified in the request.

$sel:httpStatus:ListCreatedArtifactsResponse', listCreatedArtifactsResponse_httpStatus - The response's http status code.

Response Lenses

listCreatedArtifactsResponse_nextToken :: Lens' ListCreatedArtifactsResponse (Maybe Text)

If there are more created artifacts than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

listCreatedArtifactsResponse_createdArtifactList :: Lens' ListCreatedArtifactsResponse (Maybe [CreatedArtifact])

List of created artifacts up to the maximum number of results specified in the request.

listCreatedArtifactsResponse_httpStatus :: Lens' ListCreatedArtifactsResponse Int

The response's http status code.