Network.AWS.MigrationHub.ListDiscoveredResources

Description

Lists discovered resources associated with the given MigrationTask.

This operation returns paginated results.

Creating a Request

data ListDiscoveredResources

See: newListDiscoveredResources smart constructor.

Constructors

ListDiscoveredResources'

Fields

nextToken :: Maybe Text

If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

maxResults :: Maybe Natural

The maximum number of results returned per page.

progressUpdateStream :: Text

The name of the ProgressUpdateStream.

migrationTaskName :: Text

The name of the MigrationTask. /Do not store personal data in this field./

newListDiscoveredResources

Arguments

:: Text
$sel:progressUpdateStream:ListDiscoveredResources'

-> Text
$sel:migrationTaskName:ListDiscoveredResources'

-> ListDiscoveredResources

Create a value of ListDiscoveredResources with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDiscoveredResources', listDiscoveredResources_nextToken - If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

$sel:maxResults:ListDiscoveredResources', listDiscoveredResources_maxResults - The maximum number of results returned per page.

$sel:progressUpdateStream:ListDiscoveredResources', listDiscoveredResources_progressUpdateStream - The name of the ProgressUpdateStream.

$sel:migrationTaskName:ListDiscoveredResources', listDiscoveredResources_migrationTaskName - The name of the MigrationTask. /Do not store personal data in this field./

Request Lenses

listDiscoveredResources_nextToken :: Lens' ListDiscoveredResources (Maybe Text)

If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

listDiscoveredResources_maxResults :: Lens' ListDiscoveredResources (Maybe Natural)

The maximum number of results returned per page.

listDiscoveredResources_progressUpdateStream :: Lens' ListDiscoveredResources Text

The name of the ProgressUpdateStream.

listDiscoveredResources_migrationTaskName :: Lens' ListDiscoveredResources Text

The name of the MigrationTask. /Do not store personal data in this field./

Destructuring the Response

data ListDiscoveredResourcesResponse

See: newListDiscoveredResourcesResponse smart constructor.

Constructors

ListDiscoveredResourcesResponse'

Fields

discoveredResourceList :: Maybe [DiscoveredResource]

Returned list of discovered resources associated with the given MigrationTask.

nextToken :: Maybe Text

If there are more discovered resources than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

httpStatus :: Int

The response's http status code.

newListDiscoveredResourcesResponse

Arguments

:: Int
$sel:httpStatus:ListDiscoveredResourcesResponse'

-> ListDiscoveredResourcesResponse

Create a value of ListDiscoveredResourcesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:discoveredResourceList:ListDiscoveredResourcesResponse', listDiscoveredResourcesResponse_discoveredResourceList - Returned list of discovered resources associated with the given MigrationTask.

$sel:nextToken:ListDiscoveredResources', listDiscoveredResourcesResponse_nextToken - If there are more discovered resources than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

$sel:httpStatus:ListDiscoveredResourcesResponse', listDiscoveredResourcesResponse_httpStatus - The response's http status code.

Response Lenses

listDiscoveredResourcesResponse_discoveredResourceList :: Lens' ListDiscoveredResourcesResponse (Maybe [DiscoveredResource])

Returned list of discovered resources associated with the given MigrationTask.

listDiscoveredResourcesResponse_nextToken :: Lens' ListDiscoveredResourcesResponse (Maybe Text)

If there are more discovered resources than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

listDiscoveredResourcesResponse_httpStatus :: Lens' ListDiscoveredResourcesResponse Int

The response's http status code.