Network.AWS.MigrationHub.ListMigrationTasks

Description

Lists all, or filtered by resource name, migration tasks associated with the user account making this call. This API has the following traits:

Can show a summary list of the most recent migration tasks.

Can show a summary list of migration tasks associated with a given discovered resource.

Lists migration tasks in a paginated interface.

This operation returns paginated results.

Creating a Request

data ListMigrationTasks

See: newListMigrationTasks smart constructor.

Constructors

ListMigrationTasks'

Fields

nextToken :: Maybe Text

If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

maxResults :: Maybe Natural

Value to specify how many results are returned per page.

resourceName :: Maybe Text

Filter migration tasks by discovered resource name.

newListMigrationTasks :: ListMigrationTasks

Create a value of ListMigrationTasks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListMigrationTasks', listMigrationTasks_nextToken - If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

$sel:maxResults:ListMigrationTasks', listMigrationTasks_maxResults - Value to specify how many results are returned per page.

$sel:resourceName:ListMigrationTasks', listMigrationTasks_resourceName - Filter migration tasks by discovered resource name.

Request Lenses

listMigrationTasks_nextToken :: Lens' ListMigrationTasks (Maybe Text)

If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

listMigrationTasks_maxResults :: Lens' ListMigrationTasks (Maybe Natural)

Value to specify how many results are returned per page.

listMigrationTasks_resourceName :: Lens' ListMigrationTasks (Maybe Text)

Filter migration tasks by discovered resource name.

Destructuring the Response

data ListMigrationTasksResponse

See: newListMigrationTasksResponse smart constructor.

Constructors

ListMigrationTasksResponse'

Fields

migrationTaskSummaryList :: Maybe [MigrationTaskSummary]

Lists the migration task's summary which includes: MigrationTaskName, ProgressPercent, ProgressUpdateStream, Status, and the UpdateDateTime for each task.

nextToken :: Maybe Text

If there are more migration tasks than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

httpStatus :: Int

The response's http status code.

newListMigrationTasksResponse

Arguments

:: Int
$sel:httpStatus:ListMigrationTasksResponse'

-> ListMigrationTasksResponse

Create a value of ListMigrationTasksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:migrationTaskSummaryList:ListMigrationTasksResponse', listMigrationTasksResponse_migrationTaskSummaryList - Lists the migration task's summary which includes: MigrationTaskName, ProgressPercent, ProgressUpdateStream, Status, and the UpdateDateTime for each task.

$sel:nextToken:ListMigrationTasks', listMigrationTasksResponse_nextToken - If there are more migration tasks than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

$sel:httpStatus:ListMigrationTasksResponse', listMigrationTasksResponse_httpStatus - The response's http status code.

Response Lenses

listMigrationTasksResponse_migrationTaskSummaryList :: Lens' ListMigrationTasksResponse (Maybe [MigrationTaskSummary])

Lists the migration task's summary which includes: MigrationTaskName, ProgressPercent, ProgressUpdateStream, Status, and the UpdateDateTime for each task.

listMigrationTasksResponse_nextToken :: Lens' ListMigrationTasksResponse (Maybe Text)

If there are more migration tasks than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

listMigrationTasksResponse_httpStatus :: Lens' ListMigrationTasksResponse Int

The response's http status code.