Network.AWS.MigrationHub.ListProgressUpdateStreams

Description

Lists progress update streams associated with the user account making this call.

This operation returns paginated results.

Creating a Request

data ListProgressUpdateStreams

See: newListProgressUpdateStreams smart constructor.

Constructors

ListProgressUpdateStreams'

Fields

nextToken :: Maybe Text

If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

maxResults :: Maybe Natural

Filter to limit the maximum number of results to list per page.

newListProgressUpdateStreams :: ListProgressUpdateStreams

Create a value of ListProgressUpdateStreams with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListProgressUpdateStreams', listProgressUpdateStreams_nextToken - If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

$sel:maxResults:ListProgressUpdateStreams', listProgressUpdateStreams_maxResults - Filter to limit the maximum number of results to list per page.

Request Lenses

listProgressUpdateStreams_nextToken :: Lens' ListProgressUpdateStreams (Maybe Text)

If a NextToken was returned by a previous call, there are more results available. To retrieve the next page of results, make the call again using the returned token in NextToken.

listProgressUpdateStreams_maxResults :: Lens' ListProgressUpdateStreams (Maybe Natural)

Filter to limit the maximum number of results to list per page.

Destructuring the Response

data ListProgressUpdateStreamsResponse

See: newListProgressUpdateStreamsResponse smart constructor.

Constructors

ListProgressUpdateStreamsResponse'

Fields

nextToken :: Maybe Text

If there are more streams created than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

progressUpdateStreamSummaryList :: Maybe [ProgressUpdateStreamSummary]

List of progress update streams up to the max number of results passed in the input.

httpStatus :: Int

The response's http status code.

newListProgressUpdateStreamsResponse

Arguments

:: Int
$sel:httpStatus:ListProgressUpdateStreamsResponse'

-> ListProgressUpdateStreamsResponse

Create a value of ListProgressUpdateStreamsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListProgressUpdateStreams', listProgressUpdateStreamsResponse_nextToken - If there are more streams created than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

$sel:progressUpdateStreamSummaryList:ListProgressUpdateStreamsResponse', listProgressUpdateStreamsResponse_progressUpdateStreamSummaryList - List of progress update streams up to the max number of results passed in the input.

$sel:httpStatus:ListProgressUpdateStreamsResponse', listProgressUpdateStreamsResponse_httpStatus - The response's http status code.

Response Lenses

listProgressUpdateStreamsResponse_nextToken :: Lens' ListProgressUpdateStreamsResponse (Maybe Text)

If there are more streams created than the max result, return the next token to be passed to the next call as a bookmark of where to start from.

listProgressUpdateStreamsResponse_progressUpdateStreamSummaryList :: Lens' ListProgressUpdateStreamsResponse (Maybe [ProgressUpdateStreamSummary])

List of progress update streams up to the max number of results passed in the input.

listProgressUpdateStreamsResponse_httpStatus :: Lens' ListProgressUpdateStreamsResponse Int

The response's http status code.