Network.AWS.MigrationHub.NotifyMigrationTaskState

Description

Notifies Migration Hub of the current status, progress, or other detail regarding a migration task. This API has the following traits:

Migration tools will call the NotifyMigrationTaskState API to share the latest progress and status.

MigrationTaskName is used for addressing updates to the correct target.

ProgressUpdateStream is used for access control and to provide a namespace for each migration tool.

Creating a Request

data NotifyMigrationTaskState

See: newNotifyMigrationTaskState smart constructor.

Constructors

NotifyMigrationTaskState'

Fields

dryRun :: Maybe Bool

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

progressUpdateStream :: Text

The name of the ProgressUpdateStream.

migrationTaskName :: Text

Unique identifier that references the migration task. /Do not store personal data in this field./

task :: Task

Information about the task's progress and status.

updateDateTime :: POSIX

The timestamp when the task was gathered.

nextUpdateSeconds :: Natural

Number of seconds after the UpdateDateTime within which the Migration Hub can expect an update. If Migration Hub does not receive an update within the specified interval, then the migration task will be considered stale.

newNotifyMigrationTaskState

Arguments

:: Text
$sel:progressUpdateStream:NotifyMigrationTaskState'

-> Text
$sel:migrationTaskName:NotifyMigrationTaskState'

-> Task
$sel:task:NotifyMigrationTaskState'

-> UTCTime
$sel:updateDateTime:NotifyMigrationTaskState'

-> Natural
$sel:nextUpdateSeconds:NotifyMigrationTaskState'

-> NotifyMigrationTaskState

Create a value of NotifyMigrationTaskState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:NotifyMigrationTaskState', notifyMigrationTaskState_dryRun - Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

$sel:progressUpdateStream:NotifyMigrationTaskState', notifyMigrationTaskState_progressUpdateStream - The name of the ProgressUpdateStream.

$sel:migrationTaskName:NotifyMigrationTaskState', notifyMigrationTaskState_migrationTaskName - Unique identifier that references the migration task. /Do not store personal data in this field./

$sel:task:NotifyMigrationTaskState', notifyMigrationTaskState_task - Information about the task's progress and status.

$sel:updateDateTime:NotifyMigrationTaskState', notifyMigrationTaskState_updateDateTime - The timestamp when the task was gathered.

$sel:nextUpdateSeconds:NotifyMigrationTaskState', notifyMigrationTaskState_nextUpdateSeconds - Number of seconds after the UpdateDateTime within which the Migration Hub can expect an update. If Migration Hub does not receive an update within the specified interval, then the migration task will be considered stale.

Request Lenses

notifyMigrationTaskState_dryRun :: Lens' NotifyMigrationTaskState (Maybe Bool)

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

notifyMigrationTaskState_progressUpdateStream :: Lens' NotifyMigrationTaskState Text

The name of the ProgressUpdateStream.

notifyMigrationTaskState_migrationTaskName :: Lens' NotifyMigrationTaskState Text

Unique identifier that references the migration task. /Do not store personal data in this field./

notifyMigrationTaskState_task :: Lens' NotifyMigrationTaskState Task

Information about the task's progress and status.

notifyMigrationTaskState_updateDateTime :: Lens' NotifyMigrationTaskState UTCTime

The timestamp when the task was gathered.

notifyMigrationTaskState_nextUpdateSeconds :: Lens' NotifyMigrationTaskState Natural

Number of seconds after the UpdateDateTime within which the Migration Hub can expect an update. If Migration Hub does not receive an update within the specified interval, then the migration task will be considered stale.

Destructuring the Response

data NotifyMigrationTaskStateResponse

See: newNotifyMigrationTaskStateResponse smart constructor.

Constructors

NotifyMigrationTaskStateResponse'

Fields

httpStatus :: Int

The response's http status code.

newNotifyMigrationTaskStateResponse

Arguments

:: Int
$sel:httpStatus:NotifyMigrationTaskStateResponse'

-> NotifyMigrationTaskStateResponse

Create a value of NotifyMigrationTaskStateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:NotifyMigrationTaskStateResponse', notifyMigrationTaskStateResponse_httpStatus - The response's http status code.

Response Lenses

notifyMigrationTaskStateResponse_httpStatus :: Lens' NotifyMigrationTaskStateResponse Int

The response's http status code.