Network.AWS.MigrationHub.PutResourceAttributes

Description

Provides identifying details of the resource being migrated so that it can be associated in the Application Discovery Service repository. This association occurs asynchronously after PutResourceAttributes returns.

Keep in mind that subsequent calls to PutResourceAttributes will override previously stored attributes. For example, if it is first called with a MAC address, but later, it is desired to add an IP address, it will then be required to call it with both the IP and MAC addresses to prevent overriding the MAC address.

Note the instructions regarding the special use case of the ResourceAttributeList parameter when specifying any "VM" related value.

Because this is an asynchronous call, it will always return 200, whether an association occurs or not. To confirm if an association was found based on the provided details, call ListDiscoveredResources.

Creating a Request

data PutResourceAttributes

See: newPutResourceAttributes smart constructor.

Constructors

PutResourceAttributes'

Fields

dryRun :: Maybe Bool

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

progressUpdateStream :: Text

The name of the ProgressUpdateStream.

migrationTaskName :: Text

Unique identifier that references the migration task. /Do not store personal data in this field./

resourceAttributeList :: NonEmpty ResourceAttribute

Information about the resource that is being migrated. This data will be used to map the task to a resource in the Application Discovery Service repository.

Takes the object array of ResourceAttribute where the Type field is reserved for the following values: IPV4_ADDRESS | IPV6_ADDRESS | MAC_ADDRESS | FQDN | VM_MANAGER_ID | VM_MANAGED_OBJECT_REFERENCE | VM_NAME | VM_PATH | BIOS_ID | MOTHERBOARD_SERIAL_NUMBER where the identifying value can be a string up to 256 characters.

If any "VM" related value is set for a ResourceAttribute object, it is required that VM_MANAGER_ID, as a minimum, is always set. If VM_MANAGER_ID is not set, then all "VM" fields will be discarded and "VM" fields will not be used for matching the migration task to a server in Application Discovery Service repository. See the Example section below for a use case of specifying "VM" related values.

If a server you are trying to match has multiple IP or MAC addresses, you should provide as many as you know in separate type/value pairs passed to the ResourceAttributeList parameter to maximize the chances of matching.

newPutResourceAttributes

Arguments

:: Text
$sel:progressUpdateStream:PutResourceAttributes'

-> Text
$sel:migrationTaskName:PutResourceAttributes'

-> NonEmpty ResourceAttribute
$sel:resourceAttributeList:PutResourceAttributes'

-> PutResourceAttributes

Create a value of PutResourceAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dryRun:PutResourceAttributes', putResourceAttributes_dryRun - Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

$sel:progressUpdateStream:PutResourceAttributes', putResourceAttributes_progressUpdateStream - The name of the ProgressUpdateStream.

$sel:migrationTaskName:PutResourceAttributes', putResourceAttributes_migrationTaskName - Unique identifier that references the migration task. /Do not store personal data in this field./

$sel:resourceAttributeList:PutResourceAttributes', putResourceAttributes_resourceAttributeList - Information about the resource that is being migrated. This data will be used to map the task to a resource in the Application Discovery Service repository.

Takes the object array of ResourceAttribute where the Type field is reserved for the following values: IPV4_ADDRESS | IPV6_ADDRESS | MAC_ADDRESS | FQDN | VM_MANAGER_ID | VM_MANAGED_OBJECT_REFERENCE | VM_NAME | VM_PATH | BIOS_ID | MOTHERBOARD_SERIAL_NUMBER where the identifying value can be a string up to 256 characters.

If any "VM" related value is set for a ResourceAttribute object, it is required that VM_MANAGER_ID, as a minimum, is always set. If VM_MANAGER_ID is not set, then all "VM" fields will be discarded and "VM" fields will not be used for matching the migration task to a server in Application Discovery Service repository. See the Example section below for a use case of specifying "VM" related values.

If a server you are trying to match has multiple IP or MAC addresses, you should provide as many as you know in separate type/value pairs passed to the ResourceAttributeList parameter to maximize the chances of matching.

Request Lenses

putResourceAttributes_dryRun :: Lens' PutResourceAttributes (Maybe Bool)

Optional boolean flag to indicate whether any effect should take place. Used to test if the caller has permission to make the call.

putResourceAttributes_progressUpdateStream :: Lens' PutResourceAttributes Text

The name of the ProgressUpdateStream.

putResourceAttributes_migrationTaskName :: Lens' PutResourceAttributes Text

Unique identifier that references the migration task. /Do not store personal data in this field./

putResourceAttributes_resourceAttributeList :: Lens' PutResourceAttributes (NonEmpty ResourceAttribute)

Information about the resource that is being migrated. This data will be used to map the task to a resource in the Application Discovery Service repository.

Takes the object array of ResourceAttribute where the Type field is reserved for the following values: IPV4_ADDRESS | IPV6_ADDRESS | MAC_ADDRESS | FQDN | VM_MANAGER_ID | VM_MANAGED_OBJECT_REFERENCE | VM_NAME | VM_PATH | BIOS_ID | MOTHERBOARD_SERIAL_NUMBER where the identifying value can be a string up to 256 characters.

If any "VM" related value is set for a ResourceAttribute object, it is required that VM_MANAGER_ID, as a minimum, is always set. If VM_MANAGER_ID is not set, then all "VM" fields will be discarded and "VM" fields will not be used for matching the migration task to a server in Application Discovery Service repository. See the Example section below for a use case of specifying "VM" related values.

If a server you are trying to match has multiple IP or MAC addresses, you should provide as many as you know in separate type/value pairs passed to the ResourceAttributeList parameter to maximize the chances of matching.

Destructuring the Response

data PutResourceAttributesResponse

See: newPutResourceAttributesResponse smart constructor.

Constructors

PutResourceAttributesResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutResourceAttributesResponse

Arguments

:: Int
$sel:httpStatus:PutResourceAttributesResponse'

-> PutResourceAttributesResponse

Create a value of PutResourceAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutResourceAttributesResponse', putResourceAttributesResponse_httpStatus - The response's http status code.

Response Lenses

putResourceAttributesResponse_httpStatus :: Lens' PutResourceAttributesResponse Int

The response's http status code.