Network.AWS.MigrationHub.Types

Description

Service Configuration

defaultService :: Service

API version 2017-05-31 of the Amazon Migration Hub SDK configuration.

Errors

_UnauthorizedOperation :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised to indicate a request was not authorized when the DryRun flag is set to "true".

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised when there is an internal, configuration, or dependency error encountered.

_InvalidInputException :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised when the provided input violates a policy constraint or is entered in the wrong format or data type.

_InternalServerError :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised when an internal, configuration, or dependency error is encountered.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied due to request throttling.

_PolicyErrorException :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised when there are problems accessing Application Discovery Service (Application Discovery Service); most likely due to a misconfigured policy or the migrationhub-discovery role is missing or not configured correctly.

_HomeRegionNotSetException :: AsError a => Getting (First ServiceError) a ServiceError

The home region is not set. Set the home region to continue.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised when the request references a resource (Application Discovery Service configuration, update stream, migration task, etc.) that does not exist in Application Discovery Service (Application Discovery Service) or in Migration Hub's repository.

_DryRunOperation :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised to indicate a successfully authorized action when the DryRun flag is set to "true".

ApplicationStatus

newtype ApplicationStatus

Constructors

ApplicationStatus'

Fields

fromApplicationStatus :: Text

Bundled Patterns

pattern ApplicationStatus_COMPLETED :: ApplicationStatus

pattern ApplicationStatus_IN_PROGRESS :: ApplicationStatus

pattern ApplicationStatus_NOT_STARTED :: ApplicationStatus

MigrationStatus

newtype MigrationStatus

Constructors

MigrationStatus'

Fields

fromMigrationStatus :: Text

Bundled Patterns

pattern MigrationStatus_COMPLETED :: MigrationStatus

pattern MigrationStatus_FAILED :: MigrationStatus

pattern MigrationStatus_IN_PROGRESS :: MigrationStatus

pattern MigrationStatus_NOT_STARTED :: MigrationStatus

ResourceAttributeType

newtype ResourceAttributeType

Constructors

ResourceAttributeType'

Fields

fromResourceAttributeType :: Text

Bundled Patterns

pattern ResourceAttributeType_BIOS_ID :: ResourceAttributeType

pattern ResourceAttributeType_FQDN :: ResourceAttributeType

pattern ResourceAttributeType_IPV4_ADDRESS :: ResourceAttributeType

pattern ResourceAttributeType_IPV6_ADDRESS :: ResourceAttributeType

pattern ResourceAttributeType_MAC_ADDRESS :: ResourceAttributeType

pattern ResourceAttributeType_MOTHERBOARD_SERIAL_NUMBER :: ResourceAttributeType

pattern ResourceAttributeType_VM_MANAGED_OBJECT_REFERENCE :: ResourceAttributeType

pattern ResourceAttributeType_VM_MANAGER_ID :: ResourceAttributeType

pattern ResourceAttributeType_VM_NAME :: ResourceAttributeType

pattern ResourceAttributeType_VM_PATH :: ResourceAttributeType

ApplicationState

data ApplicationState

The state of an application discovered through Migration Hub import, the AWS Agentless Discovery Connector, or the AWS Application Discovery Agent.

See: newApplicationState smart constructor.

Constructors

ApplicationState'

Fields

applicationId :: Maybe Text

The configurationId from the Application Discovery Service that uniquely identifies an application.

applicationStatus :: Maybe ApplicationStatus

The current status of an application.

lastUpdatedTime :: Maybe POSIX

The timestamp when the application status was last updated.

newApplicationState :: ApplicationState

Create a value of ApplicationState with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:ApplicationState', applicationState_applicationId - The configurationId from the Application Discovery Service that uniquely identifies an application.

$sel:applicationStatus:ApplicationState', applicationState_applicationStatus - The current status of an application.

$sel:lastUpdatedTime:ApplicationState', applicationState_lastUpdatedTime - The timestamp when the application status was last updated.

applicationState_applicationId :: Lens' ApplicationState (Maybe Text)

The configurationId from the Application Discovery Service that uniquely identifies an application.

applicationState_applicationStatus :: Lens' ApplicationState (Maybe ApplicationStatus)

The current status of an application.

applicationState_lastUpdatedTime :: Lens' ApplicationState (Maybe UTCTime)

The timestamp when the application status was last updated.

CreatedArtifact

data CreatedArtifact

An ARN of the AWS cloud resource target receiving the migration (e.g., AMI, EC2 instance, RDS instance, etc.).

See: newCreatedArtifact smart constructor.

Constructors

CreatedArtifact'

Fields

description :: Maybe Text

A description that can be free-form text to record additional detail about the artifact for clarity or for later reference.

name :: Text

An ARN that uniquely identifies the result of a migration task.

newCreatedArtifact

Arguments

:: Text
$sel:name:CreatedArtifact'

-> CreatedArtifact

Create a value of CreatedArtifact with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreatedArtifact', createdArtifact_description - A description that can be free-form text to record additional detail about the artifact for clarity or for later reference.

$sel:name:CreatedArtifact', createdArtifact_name - An ARN that uniquely identifies the result of a migration task.

createdArtifact_description :: Lens' CreatedArtifact (Maybe Text)

A description that can be free-form text to record additional detail about the artifact for clarity or for later reference.

createdArtifact_name :: Lens' CreatedArtifact Text

An ARN that uniquely identifies the result of a migration task.

DiscoveredResource

data DiscoveredResource

Object representing the on-premises resource being migrated.

See: newDiscoveredResource smart constructor.

Constructors

DiscoveredResource'

Fields

description :: Maybe Text

A description that can be free-form text to record additional detail about the discovered resource for clarity or later reference.

configurationId :: Text

The configurationId in Application Discovery Service that uniquely identifies the on-premise resource.

newDiscoveredResource

Arguments

:: Text
$sel:configurationId:DiscoveredResource'

-> DiscoveredResource

Create a value of DiscoveredResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:DiscoveredResource', discoveredResource_description - A description that can be free-form text to record additional detail about the discovered resource for clarity or later reference.

$sel:configurationId:DiscoveredResource', discoveredResource_configurationId - The configurationId in Application Discovery Service that uniquely identifies the on-premise resource.

discoveredResource_description :: Lens' DiscoveredResource (Maybe Text)

A description that can be free-form text to record additional detail about the discovered resource for clarity or later reference.

discoveredResource_configurationId :: Lens' DiscoveredResource Text

The configurationId in Application Discovery Service that uniquely identifies the on-premise resource.

MigrationTask

data MigrationTask

Represents a migration task in a migration tool.

See: newMigrationTask smart constructor.

Constructors

MigrationTask'

Fields

resourceAttributeList :: Maybe [ResourceAttribute]

Information about the resource that is being migrated. This data will be used to map the task to a resource in the Application Discovery Service repository.

updateDateTime :: Maybe POSIX

The timestamp when the task was gathered.

task :: Maybe Task

Task object encapsulating task information.

migrationTaskName :: Maybe Text

Unique identifier that references the migration task. /Do not store personal data in this field./

progressUpdateStream :: Maybe Text

A name that identifies the vendor of the migration tool being used.

newMigrationTask :: MigrationTask

Create a value of MigrationTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceAttributeList:MigrationTask', migrationTask_resourceAttributeList - Information about the resource that is being migrated. This data will be used to map the task to a resource in the Application Discovery Service repository.

$sel:updateDateTime:MigrationTask', migrationTask_updateDateTime - The timestamp when the task was gathered.

$sel:task:MigrationTask', migrationTask_task - Task object encapsulating task information.

$sel:migrationTaskName:MigrationTask', migrationTask_migrationTaskName - Unique identifier that references the migration task. /Do not store personal data in this field./

$sel:progressUpdateStream:MigrationTask', migrationTask_progressUpdateStream - A name that identifies the vendor of the migration tool being used.

migrationTask_resourceAttributeList :: Lens' MigrationTask (Maybe [ResourceAttribute])

Information about the resource that is being migrated. This data will be used to map the task to a resource in the Application Discovery Service repository.

migrationTask_updateDateTime :: Lens' MigrationTask (Maybe UTCTime)

The timestamp when the task was gathered.

migrationTask_task :: Lens' MigrationTask (Maybe Task)

Task object encapsulating task information.

migrationTask_migrationTaskName :: Lens' MigrationTask (Maybe Text)

Unique identifier that references the migration task. /Do not store personal data in this field./

migrationTask_progressUpdateStream :: Lens' MigrationTask (Maybe Text)

A name that identifies the vendor of the migration tool being used.

MigrationTaskSummary

data MigrationTaskSummary

MigrationTaskSummary includes MigrationTaskName, ProgressPercent, ProgressUpdateStream, Status, and UpdateDateTime for each task.

See: newMigrationTaskSummary smart constructor.

Constructors

MigrationTaskSummary'

Fields

status :: Maybe MigrationStatus

Status of the task.

progressPercent :: Maybe Natural

Indication of the percentage completion of the task.

updateDateTime :: Maybe POSIX

The timestamp when the task was gathered.

statusDetail :: Maybe Text

Detail information of what is being done within the overall status state.

migrationTaskName :: Maybe Text

Unique identifier that references the migration task. /Do not store personal data in this field./

progressUpdateStream :: Maybe Text

An AWS resource used for access control. It should uniquely identify the migration tool as it is used for all updates made by the tool.

newMigrationTaskSummary :: MigrationTaskSummary

Create a value of MigrationTaskSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:MigrationTaskSummary', migrationTaskSummary_status - Status of the task.

$sel:progressPercent:MigrationTaskSummary', migrationTaskSummary_progressPercent - Indication of the percentage completion of the task.

$sel:updateDateTime:MigrationTaskSummary', migrationTaskSummary_updateDateTime - The timestamp when the task was gathered.

$sel:statusDetail:MigrationTaskSummary', migrationTaskSummary_statusDetail - Detail information of what is being done within the overall status state.

$sel:migrationTaskName:MigrationTaskSummary', migrationTaskSummary_migrationTaskName - Unique identifier that references the migration task. /Do not store personal data in this field./

$sel:progressUpdateStream:MigrationTaskSummary', migrationTaskSummary_progressUpdateStream - An AWS resource used for access control. It should uniquely identify the migration tool as it is used for all updates made by the tool.

migrationTaskSummary_status :: Lens' MigrationTaskSummary (Maybe MigrationStatus)

Status of the task.

migrationTaskSummary_progressPercent :: Lens' MigrationTaskSummary (Maybe Natural)

Indication of the percentage completion of the task.

migrationTaskSummary_updateDateTime :: Lens' MigrationTaskSummary (Maybe UTCTime)

The timestamp when the task was gathered.

migrationTaskSummary_statusDetail :: Lens' MigrationTaskSummary (Maybe Text)

Detail information of what is being done within the overall status state.

migrationTaskSummary_migrationTaskName :: Lens' MigrationTaskSummary (Maybe Text)

Unique identifier that references the migration task. /Do not store personal data in this field./

migrationTaskSummary_progressUpdateStream :: Lens' MigrationTaskSummary (Maybe Text)

An AWS resource used for access control. It should uniquely identify the migration tool as it is used for all updates made by the tool.

ProgressUpdateStreamSummary

data ProgressUpdateStreamSummary

Summary of the AWS resource used for access control that is implicitly linked to your AWS account.

See: newProgressUpdateStreamSummary smart constructor.

Constructors

ProgressUpdateStreamSummary'

Fields

progressUpdateStreamName :: Maybe Text

The name of the ProgressUpdateStream. /Do not store personal data in this field./

newProgressUpdateStreamSummary :: ProgressUpdateStreamSummary

Create a value of ProgressUpdateStreamSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:progressUpdateStreamName:ProgressUpdateStreamSummary', progressUpdateStreamSummary_progressUpdateStreamName - The name of the ProgressUpdateStream. /Do not store personal data in this field./

progressUpdateStreamSummary_progressUpdateStreamName :: Lens' ProgressUpdateStreamSummary (Maybe Text)

The name of the ProgressUpdateStream. /Do not store personal data in this field./

ResourceAttribute

data ResourceAttribute

Attribute associated with a resource.

Note the corresponding format required per type listed below:

IPV4

x.x.x.x
where x is an integer in the range [0,255]

IPV6

y : y : y : y : y : y : y : y
where y is a hexadecimal between 0 and FFFF. [0, FFFF]

MAC_ADDRESS

^([0-9A-Fa-f]{2}[:-]){5}([0-9A-Fa-f]{2})$

FQDN

^[^<>{}\\\\/?,=\\p{Cntrl}]{1,256}$

See: newResourceAttribute smart constructor.

Constructors

ResourceAttribute'

Fields

type' :: ResourceAttributeType

Type of resource.

value :: Text

Value of the resource type.

newResourceAttribute

Arguments

:: ResourceAttributeType
$sel:type':ResourceAttribute'

-> Text
$sel:value:ResourceAttribute'

-> ResourceAttribute

Create a value of ResourceAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':ResourceAttribute', resourceAttribute_type - Type of resource.

$sel:value:ResourceAttribute', resourceAttribute_value - Value of the resource type.

resourceAttribute_type :: Lens' ResourceAttribute ResourceAttributeType

Type of resource.

resourceAttribute_value :: Lens' ResourceAttribute Text

Value of the resource type.

Task

data Task

Task object encapsulating task information.

See: newTask smart constructor.

Constructors

Task'

Fields

progressPercent :: Maybe Natural

Indication of the percentage completion of the task.

statusDetail :: Maybe Text

Details of task status as notified by a migration tool. A tool might use this field to provide clarifying information about the status that is unique to that tool or that explains an error state.

status :: MigrationStatus

Status of the task - Not Started, In-Progress, Complete.

newTask

Arguments

:: MigrationStatus
$sel:status:Task'

-> Task

Create a value of Task with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:progressPercent:Task', task_progressPercent - Indication of the percentage completion of the task.

$sel:statusDetail:Task', task_statusDetail - Details of task status as notified by a migration tool. A tool might use this field to provide clarifying information about the status that is unique to that tool or that explains an error state.

$sel:status:Task', task_status - Status of the task - Not Started, In-Progress, Complete.

task_progressPercent :: Lens' Task (Maybe Natural)

Indication of the percentage completion of the task.

task_statusDetail :: Lens' Task (Maybe Text)

Details of task status as notified by a migration tool. A tool might use this field to provide clarifying information about the status that is unique to that tool or that explains an error state.

task_status :: Lens' Task MigrationStatus

Status of the task - Not Started, In-Progress, Complete.