Network.AWS.MigrationHub.Types.MigrationTaskSummary

Description

Documentation

data MigrationTaskSummary

MigrationTaskSummary includes MigrationTaskName, ProgressPercent, ProgressUpdateStream, Status, and UpdateDateTime for each task.

See: newMigrationTaskSummary smart constructor.

Constructors

MigrationTaskSummary'

Fields

status :: Maybe MigrationStatus

Status of the task.

progressPercent :: Maybe Natural

Indication of the percentage completion of the task.

updateDateTime :: Maybe POSIX

The timestamp when the task was gathered.

statusDetail :: Maybe Text

Detail information of what is being done within the overall status state.

migrationTaskName :: Maybe Text

Unique identifier that references the migration task. /Do not store personal data in this field./

progressUpdateStream :: Maybe Text

An AWS resource used for access control. It should uniquely identify the migration tool as it is used for all updates made by the tool.

newMigrationTaskSummary :: MigrationTaskSummary

Create a value of MigrationTaskSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:MigrationTaskSummary', migrationTaskSummary_status - Status of the task.

$sel:progressPercent:MigrationTaskSummary', migrationTaskSummary_progressPercent - Indication of the percentage completion of the task.

$sel:updateDateTime:MigrationTaskSummary', migrationTaskSummary_updateDateTime - The timestamp when the task was gathered.

$sel:statusDetail:MigrationTaskSummary', migrationTaskSummary_statusDetail - Detail information of what is being done within the overall status state.

$sel:migrationTaskName:MigrationTaskSummary', migrationTaskSummary_migrationTaskName - Unique identifier that references the migration task. /Do not store personal data in this field./

$sel:progressUpdateStream:MigrationTaskSummary', migrationTaskSummary_progressUpdateStream - An AWS resource used for access control. It should uniquely identify the migration tool as it is used for all updates made by the tool.

migrationTaskSummary_status :: Lens' MigrationTaskSummary (Maybe MigrationStatus)

Status of the task.

migrationTaskSummary_progressPercent :: Lens' MigrationTaskSummary (Maybe Natural)

Indication of the percentage completion of the task.

migrationTaskSummary_updateDateTime :: Lens' MigrationTaskSummary (Maybe UTCTime)

The timestamp when the task was gathered.

migrationTaskSummary_statusDetail :: Lens' MigrationTaskSummary (Maybe Text)

Detail information of what is being done within the overall status state.

migrationTaskSummary_migrationTaskName :: Lens' MigrationTaskSummary (Maybe Text)

Unique identifier that references the migration task. /Do not store personal data in this field./

migrationTaskSummary_progressUpdateStream :: Lens' MigrationTaskSummary (Maybe Text)

An AWS resource used for access control. It should uniquely identify the migration tool as it is used for all updates made by the tool.