Network.AWS.MigrationHub.Types.Task

Description

Documentation

data Task

Task object encapsulating task information.

See: newTask smart constructor.

Constructors

Task'

Fields

progressPercent :: Maybe Natural

Indication of the percentage completion of the task.

statusDetail :: Maybe Text

Details of task status as notified by a migration tool. A tool might use this field to provide clarifying information about the status that is unique to that tool or that explains an error state.

status :: MigrationStatus

Status of the task - Not Started, In-Progress, Complete.

newTask

Arguments

:: MigrationStatus
$sel:status:Task'

-> Task

Create a value of Task with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:progressPercent:Task', task_progressPercent - Indication of the percentage completion of the task.

$sel:statusDetail:Task', task_statusDetail - Details of task status as notified by a migration tool. A tool might use this field to provide clarifying information about the status that is unique to that tool or that explains an error state.

$sel:status:Task', task_status - Status of the task - Not Started, In-Progress, Complete.

task_progressPercent :: Lens' Task (Maybe Natural)

Indication of the percentage completion of the task.

task_statusDetail :: Lens' Task (Maybe Text)

Details of task status as notified by a migration tool. A tool might use this field to provide clarifying information about the status that is unique to that tool or that explains an error state.

task_status :: Lens' Task MigrationStatus

Status of the task - Not Started, In-Progress, Complete.