Network.AWS.MigrationHubConfig.Lens

Description

Operations

GetHomeRegion

getHomeRegionResponse_homeRegion :: Lens' GetHomeRegionResponse (Maybe Text)

The name of the home region of the calling account.

getHomeRegionResponse_httpStatus :: Lens' GetHomeRegionResponse Int

The response's http status code.

CreateHomeRegionControl

createHomeRegionControl_dryRun :: Lens' CreateHomeRegionControl (Maybe Bool)

Optional Boolean flag to indicate whether any effect should take place. It tests whether the caller has permission to make the call.

createHomeRegionControl_homeRegion :: Lens' CreateHomeRegionControl Text

The name of the home region of the calling account.

createHomeRegionControl_target :: Lens' CreateHomeRegionControl Target

The account for which this command sets up a home region control. The Target is always of type ACCOUNT.

createHomeRegionControlResponse_homeRegionControl :: Lens' CreateHomeRegionControlResponse (Maybe HomeRegionControl)

This object is the HomeRegionControl object that's returned by a successful call to CreateHomeRegionControl.

createHomeRegionControlResponse_httpStatus :: Lens' CreateHomeRegionControlResponse Int

The response's http status code.

DescribeHomeRegionControls

describeHomeRegionControls_nextToken :: Lens' DescribeHomeRegionControls (Maybe Text)

If a NextToken was returned by a previous call, more results are available. To retrieve the next page of results, make the call again using the returned token in NextToken.

describeHomeRegionControls_homeRegion :: Lens' DescribeHomeRegionControls (Maybe Text)

The name of the home region you'd like to view.

describeHomeRegionControls_controlId :: Lens' DescribeHomeRegionControls (Maybe Text)

The ControlID is a unique identifier string of your HomeRegionControl object.

describeHomeRegionControls_maxResults :: Lens' DescribeHomeRegionControls (Maybe Natural)

The maximum number of filtering results to display per page.

describeHomeRegionControls_target :: Lens' DescribeHomeRegionControls (Maybe Target)

The target parameter specifies the identifier to which the home region is applied, which is always of type ACCOUNT. It applies the home region to the current ACCOUNT.

describeHomeRegionControlsResponse_nextToken :: Lens' DescribeHomeRegionControlsResponse (Maybe Text)

If a NextToken was returned by a previous call, more results are available. To retrieve the next page of results, make the call again using the returned token in NextToken.

describeHomeRegionControlsResponse_homeRegionControls :: Lens' DescribeHomeRegionControlsResponse (Maybe [HomeRegionControl])

An array that contains your HomeRegionControl objects.

describeHomeRegionControlsResponse_httpStatus :: Lens' DescribeHomeRegionControlsResponse Int

The response's http status code.

Types

HomeRegionControl

homeRegionControl_requestedTime :: Lens' HomeRegionControl (Maybe UTCTime)

A timestamp representing the time when the customer called CreateHomeregionControl and set the home region for the account.

homeRegionControl_homeRegion :: Lens' HomeRegionControl (Maybe Text)

The AWS Region that's been set as home region. For example, "us-west-2" or "eu-central-1" are valid home regions.

homeRegionControl_controlId :: Lens' HomeRegionControl (Maybe Text)

A unique identifier that's generated for each home region control. It's always a string that begins with "hrc-" followed by 12 lowercase letters and numbers.

homeRegionControl_target :: Lens' HomeRegionControl (Maybe Target)

The target parameter specifies the identifier to which the home region is applied, which is always an ACCOUNT. It applies the home region to the current ACCOUNT.

Target

target_id :: Lens' Target (Maybe Text)

The TargetID is a 12-character identifier of the ACCOUNT for which the control was created. (This must be the current account.)

target_type :: Lens' Target TargetType

The target type is always an ACCOUNT.