Network.AWS.MigrationHubConfig.Types

Description

Service Configuration

defaultService :: Service

API version 2019-06-30 of the Amazon Migration Hub Config SDK configuration.

Errors

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have sufficient access to perform this action.

_DryRunOperation :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised to indicate that authorization of an action was successful, when the DryRun flag is set to true.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The request was denied due to request throttling.

_InternalServerError :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised when an internal, configuration, or dependency error is encountered.

_InvalidInputException :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised when the provided input violates a policy constraint or is entered in the wrong format or data type.

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

Exception raised when a request fails due to temporary unavailability of the service.

TargetType

newtype TargetType

Constructors

TargetType'

Fields

fromTargetType :: Text

Bundled Patterns

pattern TargetType_ACCOUNT :: TargetType

HomeRegionControl

data HomeRegionControl

A home region control is an object that specifies the home region for an account, with some additional information. It contains a target (always of type ACCOUNT), an ID, and a time at which the home region was set.

See: newHomeRegionControl smart constructor.

Constructors

HomeRegionControl'

Fields

requestedTime :: Maybe POSIX

A timestamp representing the time when the customer called CreateHomeregionControl and set the home region for the account.

homeRegion :: Maybe Text

The AWS Region that's been set as home region. For example, "us-west-2" or "eu-central-1" are valid home regions.

controlId :: Maybe Text

A unique identifier that's generated for each home region control. It's always a string that begins with "hrc-" followed by 12 lowercase letters and numbers.

target :: Maybe Target

The target parameter specifies the identifier to which the home region is applied, which is always an ACCOUNT. It applies the home region to the current ACCOUNT.

newHomeRegionControl :: HomeRegionControl

Create a value of HomeRegionControl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestedTime:HomeRegionControl', homeRegionControl_requestedTime - A timestamp representing the time when the customer called CreateHomeregionControl and set the home region for the account.

$sel:homeRegion:HomeRegionControl', homeRegionControl_homeRegion - The AWS Region that's been set as home region. For example, "us-west-2" or "eu-central-1" are valid home regions.

$sel:controlId:HomeRegionControl', homeRegionControl_controlId - A unique identifier that's generated for each home region control. It's always a string that begins with "hrc-" followed by 12 lowercase letters and numbers.

$sel:target:HomeRegionControl', homeRegionControl_target - The target parameter specifies the identifier to which the home region is applied, which is always an ACCOUNT. It applies the home region to the current ACCOUNT.

homeRegionControl_requestedTime :: Lens' HomeRegionControl (Maybe UTCTime)

A timestamp representing the time when the customer called CreateHomeregionControl and set the home region for the account.

homeRegionControl_homeRegion :: Lens' HomeRegionControl (Maybe Text)

The AWS Region that's been set as home region. For example, "us-west-2" or "eu-central-1" are valid home regions.

homeRegionControl_controlId :: Lens' HomeRegionControl (Maybe Text)

A unique identifier that's generated for each home region control. It's always a string that begins with "hrc-" followed by 12 lowercase letters and numbers.

homeRegionControl_target :: Lens' HomeRegionControl (Maybe Target)

The target parameter specifies the identifier to which the home region is applied, which is always an ACCOUNT. It applies the home region to the current ACCOUNT.

Target

data Target

The target parameter specifies the identifier to which the home region is applied, which is always an ACCOUNT. It applies the home region to the current ACCOUNT.

See: newTarget smart constructor.

Constructors

Target'

Fields

id :: Maybe Text

The TargetID is a 12-character identifier of the ACCOUNT for which the control was created. (This must be the current account.)

type' :: TargetType

The target type is always an ACCOUNT.

newTarget

Arguments

:: TargetType
$sel:type':Target'

-> Target

Create a value of Target with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:Target', target_id - The TargetID is a 12-character identifier of the ACCOUNT for which the control was created. (This must be the current account.)

$sel:type':Target', target_type - The target type is always an ACCOUNT.

target_id :: Lens' Target (Maybe Text)

The TargetID is a 12-character identifier of the ACCOUNT for which the control was created. (This must be the current account.)

target_type :: Lens' Target TargetType

The target type is always an ACCOUNT.