Network.AWS.Mobile

Description

Derived from API version 2017-07-01 of the AWS service descriptions, licensed under Apache 2.0.

AWS Mobile Service provides mobile app and website developers with capabilities required to configure AWS resources and bootstrap their developer desktop projects with the necessary SDKs, constants, tools and samples to make use of those resources.

Service Configuration

defaultService :: Service

API version 2017-07-01 of the Amazon Mobile SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by Mobile.

NotFoundException

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

No entity can be found with the specified identifier.

BadRequestException

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

The request cannot be processed because some parameter is not valid or the project state prevents the operation from being performed.

UnauthorizedException

_UnauthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

Credentials of the caller are insufficient to authorize the request.

ServiceUnavailableException

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The service is temporarily unavailable. The request should be retried after some time delay.

LimitExceededException

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

There are too many AWS Mobile Hub projects in the account or the account has exceeded the maximum number of resources in some AWS service. You should create another sub-account using AWS Organizations or remove some resources and retry your request.

AccountActionRequiredException

_AccountActionRequiredException :: AsError a => Getting (First ServiceError) a ServiceError

Account Action is required in order to continue the request.

InternalFailureException

_InternalFailureException :: AsError a => Getting (First ServiceError) a ServiceError

The service has encountered an unexpected error condition which prevents it from servicing the request.

TooManyRequestsException

_TooManyRequestsException :: AsError a => Getting (First ServiceError) a ServiceError

Too many requests have been received for this AWS account in too short a time. The request should be retried after some time delay.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

ListProjects (Paginated)

data ListProjects

Request structure used to request projects list in AWS Mobile Hub.

See: newListProjects smart constructor.

Constructors

ListProjects' (Maybe Text) (Maybe Int)

newListProjects :: ListProjects

Create a value of ListProjects with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListProjects', listProjects_nextToken - Pagination token. Set to null to start listing projects from start. If non-null pagination token is returned in a result, then pass its value in here in another request to list more projects.

$sel:maxResults:ListProjects', listProjects_maxResults - Maximum number of records to list in a single response.

data ListProjectsResponse

Result structure used for requests to list projects in AWS Mobile Hub.

See: newListProjectsResponse smart constructor.

Constructors

ListProjectsResponse' (Maybe Text) (Maybe [ProjectSummary]) Int

newListProjectsResponse

Arguments

:: Int
$sel:httpStatus:ListProjectsResponse'

-> ListProjectsResponse

Create a value of ListProjectsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListProjects', listProjectsResponse_nextToken - Undocumented member.

$sel:projects:ListProjectsResponse', listProjectsResponse_projects - Undocumented member.

$sel:httpStatus:ListProjectsResponse', listProjectsResponse_httpStatus - The response's http status code.

CreateProject

data CreateProject

Request structure used to request a project be created.

See: newCreateProject smart constructor.

Constructors

CreateProject' (Maybe ByteString) (Maybe Text) (Maybe Text) (Maybe Text)

newCreateProject :: CreateProject

Create a value of CreateProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contents:CreateProject', createProject_contents - ZIP or YAML file which contains configuration settings to be used when creating the project. This may be the contents of the file downloaded from the URL provided in an export project operation.

$sel:name:CreateProject', createProject_name - Name of the project.

$sel:snapshotId:CreateProject', createProject_snapshotId - Unique identifier for an exported snapshot of project configuration. This snapshot identifier is included in the share URL when a project is exported.

$sel:region:CreateProject', createProject_region - Default region where project resources should be created.

data CreateProjectResponse

Result structure used in response to a request to create a project.

See: newCreateProjectResponse smart constructor.

Constructors

CreateProjectResponse' (Maybe ProjectDetails) Int

newCreateProjectResponse

Arguments

:: Int
$sel:httpStatus:CreateProjectResponse'

-> CreateProjectResponse

Create a value of CreateProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:details:CreateProjectResponse', createProjectResponse_details - Detailed information about the created AWS Mobile Hub project.

$sel:httpStatus:CreateProjectResponse', createProjectResponse_httpStatus - The response's http status code.

DescribeBundle

data DescribeBundle

Request structure to request the details of a specific bundle.

See: newDescribeBundle smart constructor.

Constructors

DescribeBundle' Text

newDescribeBundle

Arguments

:: Text
$sel:bundleId:DescribeBundle'

-> DescribeBundle

Create a value of DescribeBundle with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bundleId:DescribeBundle', describeBundle_bundleId - Unique bundle identifier.

data DescribeBundleResponse

Result structure contains the details of the bundle.

See: newDescribeBundleResponse smart constructor.

Constructors

DescribeBundleResponse' (Maybe BundleDetails) Int

newDescribeBundleResponse

Arguments

:: Int
$sel:httpStatus:DescribeBundleResponse'

-> DescribeBundleResponse

Create a value of DescribeBundleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:details:DescribeBundleResponse', describeBundleResponse_details - The details of the bundle.

$sel:httpStatus:DescribeBundleResponse', describeBundleResponse_httpStatus - The response's http status code.

ListBundles (Paginated)

data ListBundles

Request structure to request all available bundles.

See: newListBundles smart constructor.

Constructors

ListBundles' (Maybe Text) (Maybe Int)

newListBundles :: ListBundles

Create a value of ListBundles with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListBundles', listBundles_nextToken - Pagination token. Set to null to start listing bundles from start. If non-null pagination token is returned in a result, then pass its value in here in another request to list more bundles.

$sel:maxResults:ListBundles', listBundles_maxResults - Maximum number of records to list in a single response.

data ListBundlesResponse

Result structure contains a list of all available bundles with details.

See: newListBundlesResponse smart constructor.

Constructors

ListBundlesResponse' (Maybe Text) (Maybe [BundleDetails]) Int

newListBundlesResponse

Arguments

:: Int
$sel:httpStatus:ListBundlesResponse'

-> ListBundlesResponse

Create a value of ListBundlesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListBundles', listBundlesResponse_nextToken - Pagination token. If non-null pagination token is returned in a result, then pass its value in another request to fetch more entries.

$sel:bundleList:ListBundlesResponse', listBundlesResponse_bundleList - A list of bundles.

$sel:httpStatus:ListBundlesResponse', listBundlesResponse_httpStatus - The response's http status code.

UpdateProject

data UpdateProject

Request structure used for requests to update project configuration.

See: newUpdateProject smart constructor.

Constructors

UpdateProject' (Maybe ByteString) Text

newUpdateProject

Arguments

:: Text
$sel:projectId:UpdateProject'

-> UpdateProject

Create a value of UpdateProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:contents:UpdateProject', updateProject_contents - ZIP or YAML file which contains project configuration to be updated. This should be the contents of the file downloaded from the URL provided in an export project operation.

$sel:projectId:UpdateProject', updateProject_projectId - Unique project identifier.

data UpdateProjectResponse

Result structure used for requests to updated project configuration.

See: newUpdateProjectResponse smart constructor.

Constructors

UpdateProjectResponse' (Maybe ProjectDetails) Int

newUpdateProjectResponse

Arguments

:: Int
$sel:httpStatus:UpdateProjectResponse'

-> UpdateProjectResponse

Create a value of UpdateProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:details:UpdateProjectResponse', updateProjectResponse_details - Detailed information about the updated AWS Mobile Hub project.

$sel:httpStatus:UpdateProjectResponse', updateProjectResponse_httpStatus - The response's http status code.

DeleteProject

data DeleteProject

Request structure used to request a project be deleted.

See: newDeleteProject smart constructor.

Constructors

DeleteProject' Text

newDeleteProject

Arguments

:: Text
$sel:projectId:DeleteProject'

-> DeleteProject

Create a value of DeleteProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectId:DeleteProject', deleteProject_projectId - Unique project identifier.

data DeleteProjectResponse

Result structure used in response to request to delete a project.

See: newDeleteProjectResponse smart constructor.

Constructors

DeleteProjectResponse' (Maybe [Resource]) (Maybe [Resource]) Int

newDeleteProjectResponse

Arguments

:: Int
$sel:httpStatus:DeleteProjectResponse'

-> DeleteProjectResponse

Create a value of DeleteProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deletedResources:DeleteProjectResponse', deleteProjectResponse_deletedResources - Resources which were deleted.

$sel:orphanedResources:DeleteProjectResponse', deleteProjectResponse_orphanedResources - Resources which were not deleted, due to a risk of losing potentially important data or files.

$sel:httpStatus:DeleteProjectResponse', deleteProjectResponse_httpStatus - The response's http status code.

ExportProject

data ExportProject

Request structure used in requests to export project configuration details.

See: newExportProject smart constructor.

Constructors

ExportProject' Text

newExportProject

Arguments

:: Text
$sel:projectId:ExportProject'

-> ExportProject

Create a value of ExportProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectId:ExportProject', exportProject_projectId - Unique project identifier.

data ExportProjectResponse

Result structure used for requests to export project configuration details.

See: newExportProjectResponse smart constructor.

Constructors

ExportProjectResponse' (Maybe Text) (Maybe Text) (Maybe Text) Int

newExportProjectResponse

Arguments

:: Int
$sel:httpStatus:ExportProjectResponse'

-> ExportProjectResponse

Create a value of ExportProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:downloadUrl:ExportProjectResponse', exportProjectResponse_downloadUrl - URL which can be used to download the exported project configuation file(s).

$sel:snapshotId:ExportProjectResponse', exportProjectResponse_snapshotId - Unique identifier for the exported snapshot of the project configuration. This snapshot identifier is included in the share URL.

$sel:shareUrl:ExportProjectResponse', exportProjectResponse_shareUrl - URL which can be shared to allow other AWS users to create their own project in AWS Mobile Hub with the same configuration as the specified project. This URL pertains to a snapshot in time of the project configuration that is created when this API is called. If you want to share additional changes to your project configuration, then you will need to create and share a new snapshot by calling this method again.

$sel:httpStatus:ExportProjectResponse', exportProjectResponse_httpStatus - The response's http status code.

DescribeProject

data DescribeProject

Request structure used to request details about a project.

See: newDescribeProject smart constructor.

Constructors

DescribeProject' (Maybe Bool) Text

newDescribeProject

Arguments

:: Text
$sel:projectId:DescribeProject'

-> DescribeProject

Create a value of DescribeProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:syncFromResources:DescribeProject', describeProject_syncFromResources - If set to true, causes AWS Mobile Hub to synchronize information from other services, e.g., update state of AWS CloudFormation stacks in the AWS Mobile Hub project.

$sel:projectId:DescribeProject', describeProject_projectId - Unique project identifier.

data DescribeProjectResponse

Result structure used for requests of project details.

See: newDescribeProjectResponse smart constructor.

Constructors

DescribeProjectResponse' (Maybe ProjectDetails) Int

newDescribeProjectResponse

Arguments

:: Int
$sel:httpStatus:DescribeProjectResponse'

-> DescribeProjectResponse

Create a value of DescribeProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:details:DescribeProjectResponse', describeProjectResponse_details - Undocumented member.

$sel:httpStatus:DescribeProjectResponse', describeProjectResponse_httpStatus - The response's http status code.

ExportBundle

data ExportBundle

Request structure used to request generation of custom SDK and tool packages required to integrate mobile web or app clients with backed AWS resources.

See: newExportBundle smart constructor.

Constructors

ExportBundle' (Maybe Platform) (Maybe Text) Text

newExportBundle

Arguments

:: Text
$sel:bundleId:ExportBundle'

-> ExportBundle

Create a value of ExportBundle with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platform:ExportBundle', exportBundle_platform - Developer desktop or target application platform.

$sel:projectId:ExportBundle', exportBundle_projectId - Unique project identifier.

$sel:bundleId:ExportBundle', exportBundle_bundleId - Unique bundle identifier.

data ExportBundleResponse

Result structure which contains link to download custom-generated SDK and tool packages used to integrate mobile web or app clients with backed AWS resources.

See: newExportBundleResponse smart constructor.

Constructors

ExportBundleResponse' (Maybe Text) Int

newExportBundleResponse

Arguments

:: Int
$sel:httpStatus:ExportBundleResponse'

-> ExportBundleResponse

Create a value of ExportBundleResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:downloadUrl:ExportBundleResponse', exportBundleResponse_downloadUrl - URL which contains the custom-generated SDK and tool packages used to integrate the client mobile app or web app with the AWS resources created by the AWS Mobile Hub project.

$sel:httpStatus:ExportBundleResponse', exportBundleResponse_httpStatus - The response's http status code.

Types

Platform

newtype Platform

Developer desktop or target mobile app or website platform.

Constructors

Platform'

Fields

fromPlatform :: Text

Bundled Patterns

pattern Platform_ANDROID :: Platform

pattern Platform_JAVASCRIPT :: Platform

pattern Platform_LINUX :: Platform

pattern Platform_OBJC :: Platform

pattern Platform_OSX :: Platform

pattern Platform_SWIFT :: Platform

pattern Platform_WINDOWS :: Platform

ProjectState

newtype ProjectState

Synchronization state for a project.

Constructors

ProjectState'

Fields

fromProjectState :: Text

Bundled Patterns

pattern ProjectState_IMPORTING :: ProjectState

pattern ProjectState_NORMAL :: ProjectState

pattern ProjectState_SYNCING :: ProjectState

BundleDetails

data BundleDetails

The details of the bundle.

See: newBundleDetails smart constructor.

Constructors

BundleDetails' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe Text) (Maybe [Platform])

newBundleDetails :: BundleDetails

Create a value of BundleDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bundleId:BundleDetails', bundleDetails_bundleId - Undocumented member.

$sel:iconUrl:BundleDetails', bundleDetails_iconUrl - Undocumented member.

$sel:title:BundleDetails', bundleDetails_title - Undocumented member.

$sel:version:BundleDetails', bundleDetails_version - Undocumented member.

$sel:description:BundleDetails', bundleDetails_description - Undocumented member.

$sel:availablePlatforms:BundleDetails', bundleDetails_availablePlatforms - Undocumented member.

ProjectDetails

data ProjectDetails

Detailed information about an AWS Mobile Hub project.

See: newProjectDetails smart constructor.

Constructors

ProjectDetails' (Maybe POSIX) (Maybe POSIX) (Maybe Text) (Maybe ProjectState) (Maybe Text) (Maybe Text) (Maybe [Resource]) (Maybe Text)

newProjectDetails :: ProjectDetails

Create a value of ProjectDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdDate:ProjectDetails', projectDetails_createdDate - Date the project was created.

$sel:lastUpdatedDate:ProjectDetails', projectDetails_lastUpdatedDate - Date of the last modification of the project.

$sel:projectId:ProjectDetails', projectDetails_projectId - Undocumented member.

$sel:state:ProjectDetails', projectDetails_state - Undocumented member.

$sel:name:ProjectDetails', projectDetails_name - Undocumented member.

$sel:consoleUrl:ProjectDetails', projectDetails_consoleUrl - Website URL for this project in the AWS Mobile Hub console.

$sel:resources:ProjectDetails', projectDetails_resources - Undocumented member.

$sel:region:ProjectDetails', projectDetails_region - Undocumented member.

ProjectSummary

data ProjectSummary

Summary information about an AWS Mobile Hub project.

See: newProjectSummary smart constructor.

Constructors

ProjectSummary' (Maybe Text) (Maybe Text)

newProjectSummary :: ProjectSummary

Create a value of ProjectSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectId:ProjectSummary', projectSummary_projectId - Unique project identifier.

$sel:name:ProjectSummary', projectSummary_name - Name of the project.

Resource

data Resource

Information about an instance of an AWS resource associated with a project.

See: newResource smart constructor.

Constructors

Resource' (Maybe Text) (Maybe Text) (Maybe (HashMap Text Text)) (Maybe Text) (Maybe Text)

newResource :: Resource

Create a value of Resource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Resource', resource_arn - Undocumented member.

$sel:name:Resource', resource_name - Undocumented member.

$sel:attributes:Resource', resource_attributes - Undocumented member.

$sel:feature:Resource', resource_feature - Undocumented member.

$sel:type':Resource', resource_type - Undocumented member.