Network.AWS.Mobile.ExportProject

Description

Exports project configuration to a snapshot which can be downloaded and shared. Note that mobile app push credentials are encrypted in exported projects, so they can only be shared successfully within the same AWS account.

Creating a Request

data ExportProject

Request structure used in requests to export project configuration details.

See: newExportProject smart constructor.

Constructors

ExportProject'

Fields

projectId :: Text

Unique project identifier.

newExportProject

Arguments

:: Text
$sel:projectId:ExportProject'

-> ExportProject

Create a value of ExportProject with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:projectId:ExportProject', exportProject_projectId - Unique project identifier.

Request Lenses

exportProject_projectId :: Lens' ExportProject Text

Unique project identifier.

Destructuring the Response

data ExportProjectResponse

Result structure used for requests to export project configuration details.

See: newExportProjectResponse smart constructor.

Constructors

ExportProjectResponse'

Fields

shareUrl :: Maybe Text

URL which can be shared to allow other AWS users to create their own project in AWS Mobile Hub with the same configuration as the specified project. This URL pertains to a snapshot in time of the project configuration that is created when this API is called. If you want to share additional changes to your project configuration, then you will need to create and share a new snapshot by calling this method again.

downloadUrl :: Maybe Text

URL which can be used to download the exported project configuation file(s).

snapshotId :: Maybe Text

Unique identifier for the exported snapshot of the project configuration. This snapshot identifier is included in the share URL.

httpStatus :: Int

The response's http status code.

newExportProjectResponse

Arguments

:: Int
$sel:httpStatus:ExportProjectResponse'

-> ExportProjectResponse

Create a value of ExportProjectResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:shareUrl:ExportProjectResponse', exportProjectResponse_shareUrl - URL which can be shared to allow other AWS users to create their own project in AWS Mobile Hub with the same configuration as the specified project. This URL pertains to a snapshot in time of the project configuration that is created when this API is called. If you want to share additional changes to your project configuration, then you will need to create and share a new snapshot by calling this method again.

$sel:downloadUrl:ExportProjectResponse', exportProjectResponse_downloadUrl - URL which can be used to download the exported project configuation file(s).

$sel:snapshotId:ExportProjectResponse', exportProjectResponse_snapshotId - Unique identifier for the exported snapshot of the project configuration. This snapshot identifier is included in the share URL.

$sel:httpStatus:ExportProjectResponse', exportProjectResponse_httpStatus - The response's http status code.

Response Lenses

exportProjectResponse_shareUrl :: Lens' ExportProjectResponse (Maybe Text)

URL which can be shared to allow other AWS users to create their own project in AWS Mobile Hub with the same configuration as the specified project. This URL pertains to a snapshot in time of the project configuration that is created when this API is called. If you want to share additional changes to your project configuration, then you will need to create and share a new snapshot by calling this method again.

exportProjectResponse_downloadUrl :: Lens' ExportProjectResponse (Maybe Text)

URL which can be used to download the exported project configuation file(s).

exportProjectResponse_snapshotId :: Lens' ExportProjectResponse (Maybe Text)

Unique identifier for the exported snapshot of the project configuration. This snapshot identifier is included in the share URL.

exportProjectResponse_httpStatus :: Lens' ExportProjectResponse Int

The response's http status code.