Network.AWS.Mobile.Lens

Description

Operations

ListProjects

listProjects_nextToken :: Lens' ListProjects (Maybe Text)

Pagination token. Set to null to start listing projects from start. If non-null pagination token is returned in a result, then pass its value in here in another request to list more projects.

listProjects_maxResults :: Lens' ListProjects (Maybe Int)

Maximum number of records to list in a single response.

listProjectsResponse_nextToken :: Lens' ListProjectsResponse (Maybe Text)

Undocumented member.

listProjectsResponse_projects :: Lens' ListProjectsResponse (Maybe [ProjectSummary])

Undocumented member.

listProjectsResponse_httpStatus :: Lens' ListProjectsResponse Int

The response's http status code.

CreateProject

createProject_contents :: Lens' CreateProject (Maybe ByteString)

ZIP or YAML file which contains configuration settings to be used when creating the project. This may be the contents of the file downloaded from the URL provided in an export project operation.

createProject_name :: Lens' CreateProject (Maybe Text)

Name of the project.

createProject_snapshotId :: Lens' CreateProject (Maybe Text)

Unique identifier for an exported snapshot of project configuration. This snapshot identifier is included in the share URL when a project is exported.

createProject_region :: Lens' CreateProject (Maybe Text)

Default region where project resources should be created.

createProjectResponse_details :: Lens' CreateProjectResponse (Maybe ProjectDetails)

Detailed information about the created AWS Mobile Hub project.

createProjectResponse_httpStatus :: Lens' CreateProjectResponse Int

The response's http status code.

DescribeBundle

describeBundle_bundleId :: Lens' DescribeBundle Text

Unique bundle identifier.

describeBundleResponse_details :: Lens' DescribeBundleResponse (Maybe BundleDetails)

The details of the bundle.

describeBundleResponse_httpStatus :: Lens' DescribeBundleResponse Int

The response's http status code.

ListBundles

listBundles_nextToken :: Lens' ListBundles (Maybe Text)

Pagination token. Set to null to start listing bundles from start. If non-null pagination token is returned in a result, then pass its value in here in another request to list more bundles.

listBundles_maxResults :: Lens' ListBundles (Maybe Int)

Maximum number of records to list in a single response.

listBundlesResponse_nextToken :: Lens' ListBundlesResponse (Maybe Text)

Pagination token. If non-null pagination token is returned in a result, then pass its value in another request to fetch more entries.

listBundlesResponse_bundleList :: Lens' ListBundlesResponse (Maybe [BundleDetails])

A list of bundles.

listBundlesResponse_httpStatus :: Lens' ListBundlesResponse Int

The response's http status code.

UpdateProject

updateProject_contents :: Lens' UpdateProject (Maybe ByteString)

ZIP or YAML file which contains project configuration to be updated. This should be the contents of the file downloaded from the URL provided in an export project operation.

updateProject_projectId :: Lens' UpdateProject Text

Unique project identifier.

updateProjectResponse_details :: Lens' UpdateProjectResponse (Maybe ProjectDetails)

Detailed information about the updated AWS Mobile Hub project.

updateProjectResponse_httpStatus :: Lens' UpdateProjectResponse Int

The response's http status code.

DeleteProject

deleteProject_projectId :: Lens' DeleteProject Text

Unique project identifier.

deleteProjectResponse_deletedResources :: Lens' DeleteProjectResponse (Maybe [Resource])

Resources which were deleted.

deleteProjectResponse_orphanedResources :: Lens' DeleteProjectResponse (Maybe [Resource])

Resources which were not deleted, due to a risk of losing potentially important data or files.

deleteProjectResponse_httpStatus :: Lens' DeleteProjectResponse Int

The response's http status code.

ExportProject

exportProject_projectId :: Lens' ExportProject Text

Unique project identifier.

exportProjectResponse_downloadUrl :: Lens' ExportProjectResponse (Maybe Text)

URL which can be used to download the exported project configuation file(s).

exportProjectResponse_snapshotId :: Lens' ExportProjectResponse (Maybe Text)

Unique identifier for the exported snapshot of the project configuration. This snapshot identifier is included in the share URL.

exportProjectResponse_shareUrl :: Lens' ExportProjectResponse (Maybe Text)

URL which can be shared to allow other AWS users to create their own project in AWS Mobile Hub with the same configuration as the specified project. This URL pertains to a snapshot in time of the project configuration that is created when this API is called. If you want to share additional changes to your project configuration, then you will need to create and share a new snapshot by calling this method again.

exportProjectResponse_httpStatus :: Lens' ExportProjectResponse Int

The response's http status code.

DescribeProject

describeProject_syncFromResources :: Lens' DescribeProject (Maybe Bool)

If set to true, causes AWS Mobile Hub to synchronize information from other services, e.g., update state of AWS CloudFormation stacks in the AWS Mobile Hub project.

describeProject_projectId :: Lens' DescribeProject Text

Unique project identifier.

describeProjectResponse_details :: Lens' DescribeProjectResponse (Maybe ProjectDetails)

Undocumented member.

describeProjectResponse_httpStatus :: Lens' DescribeProjectResponse Int

The response's http status code.

ExportBundle

exportBundle_platform :: Lens' ExportBundle (Maybe Platform)

Developer desktop or target application platform.

exportBundle_projectId :: Lens' ExportBundle (Maybe Text)

Unique project identifier.

exportBundle_bundleId :: Lens' ExportBundle Text

Unique bundle identifier.

exportBundleResponse_downloadUrl :: Lens' ExportBundleResponse (Maybe Text)

URL which contains the custom-generated SDK and tool packages used to integrate the client mobile app or web app with the AWS resources created by the AWS Mobile Hub project.

exportBundleResponse_httpStatus :: Lens' ExportBundleResponse Int

The response's http status code.

Types

BundleDetails

bundleDetails_bundleId :: Lens' BundleDetails (Maybe Text)

Undocumented member.

bundleDetails_iconUrl :: Lens' BundleDetails (Maybe Text)

Undocumented member.

bundleDetails_title :: Lens' BundleDetails (Maybe Text)

Undocumented member.

bundleDetails_version :: Lens' BundleDetails (Maybe Text)

Undocumented member.

bundleDetails_description :: Lens' BundleDetails (Maybe Text)

Undocumented member.

bundleDetails_availablePlatforms :: Lens' BundleDetails (Maybe [Platform])

Undocumented member.

ProjectDetails

projectDetails_createdDate :: Lens' ProjectDetails (Maybe UTCTime)

Date the project was created.

projectDetails_lastUpdatedDate :: Lens' ProjectDetails (Maybe UTCTime)

Date of the last modification of the project.

projectDetails_projectId :: Lens' ProjectDetails (Maybe Text)

Undocumented member.

projectDetails_state :: Lens' ProjectDetails (Maybe ProjectState)

Undocumented member.

projectDetails_name :: Lens' ProjectDetails (Maybe Text)

Undocumented member.

projectDetails_consoleUrl :: Lens' ProjectDetails (Maybe Text)

Website URL for this project in the AWS Mobile Hub console.

projectDetails_resources :: Lens' ProjectDetails (Maybe [Resource])

Undocumented member.

projectDetails_region :: Lens' ProjectDetails (Maybe Text)

Undocumented member.

ProjectSummary

projectSummary_projectId :: Lens' ProjectSummary (Maybe Text)

Unique project identifier.

projectSummary_name :: Lens' ProjectSummary (Maybe Text)

Name of the project.

Resource

resource_arn :: Lens' Resource (Maybe Text)

Undocumented member.

resource_name :: Lens' Resource (Maybe Text)

Undocumented member.

resource_attributes :: Lens' Resource (Maybe (HashMap Text Text))

Undocumented member.

resource_feature :: Lens' Resource (Maybe Text)

Undocumented member.

resource_type :: Lens' Resource (Maybe Text)

Undocumented member.