Network.AWS.Mobile.Types.Resource

Description

Documentation

data Resource

Information about an instance of an AWS resource associated with a project.

See: newResource smart constructor.

Constructors

Resource'

Fields

arn :: Maybe Text

name :: Maybe Text

attributes :: Maybe (HashMap Text Text)

feature :: Maybe Text

type' :: Maybe Text

newResource :: Resource

Create a value of Resource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:Resource', resource_arn - Undocumented member.

$sel:name:Resource', resource_name - Undocumented member.

$sel:attributes:Resource', resource_attributes - Undocumented member.

$sel:feature:Resource', resource_feature - Undocumented member.

$sel:type':Resource', resource_type - Undocumented member.

resource_arn :: Lens' Resource (Maybe Text)

Undocumented member.

resource_name :: Lens' Resource (Maybe Text)

Undocumented member.

resource_attributes :: Lens' Resource (Maybe (HashMap Text Text))

Undocumented member.

resource_feature :: Lens' Resource (Maybe Text)

Undocumented member.

resource_type :: Lens' Resource (Maybe Text)

Undocumented member.