Network.AWS.MwAA.CreateCliToken

Description

Create a CLI token to use Airflow CLI.

Creating a Request

data CreateCliToken

See: newCreateCliToken smart constructor.

Constructors

CreateCliToken'

Fields

name :: Text

Create a CLI token request for a MWAA environment.

newCreateCliToken

Arguments

:: Text
$sel:name:CreateCliToken'

-> CreateCliToken

Create a value of CreateCliToken with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:CreateCliToken', createCliToken_name - Create a CLI token request for a MWAA environment.

Request Lenses

createCliToken_name :: Lens' CreateCliToken Text

Create a CLI token request for a MWAA environment.

Destructuring the Response

data CreateCliTokenResponse

See: newCreateCliTokenResponse smart constructor.

Constructors

CreateCliTokenResponse'

Fields

webServerHostname :: Maybe Text

Create an Airflow CLI login token response for the provided webserver hostname.

cliToken :: Maybe (Sensitive Text)

Create an Airflow CLI login token response for the provided JWT token.

httpStatus :: Int

The response's http status code.

newCreateCliTokenResponse

Arguments

:: Int
$sel:httpStatus:CreateCliTokenResponse'

-> CreateCliTokenResponse

Create a value of CreateCliTokenResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:webServerHostname:CreateCliTokenResponse', createCliTokenResponse_webServerHostname - Create an Airflow CLI login token response for the provided webserver hostname.

$sel:cliToken:CreateCliTokenResponse', createCliTokenResponse_cliToken - Create an Airflow CLI login token response for the provided JWT token.

$sel:httpStatus:CreateCliTokenResponse', createCliTokenResponse_httpStatus - The response's http status code.

Response Lenses

createCliTokenResponse_webServerHostname :: Lens' CreateCliTokenResponse (Maybe Text)

Create an Airflow CLI login token response for the provided webserver hostname.

createCliTokenResponse_cliToken :: Lens' CreateCliTokenResponse (Maybe Text)

Create an Airflow CLI login token response for the provided JWT token.

createCliTokenResponse_httpStatus :: Lens' CreateCliTokenResponse Int

The response's http status code.