Network.AWS.MwAA.Types.Environment

Description

Documentation

data Environment

The Amazon Managed Workflows for Apache Airflow (MWAA) environment.

See: newEnvironment smart constructor.

Constructors

Environment'

Fields

serviceRoleArn :: Maybe Text

The Amazon Resource Name (ARN) for the service-linked role of the environment. To learn more, see Amazon MWAA Service-linked role.

schedulers :: Maybe Int

The number of Apache Airflow schedulers that run in your Amazon MWAA environment.

status :: Maybe EnvironmentStatus

The status of the Amazon MWAA environment. Valid values:

CREATING - Indicates the request to create the environment is in progress.

CREATE_FAILED - Indicates the request to create the environment failed, and the environment could not be created.

AVAILABLE - Indicates the request was successful and the environment is ready to use.

UPDATING - Indicates the request to update the environment is in progress.

DELETING - Indicates the request to delete the environment is in progress.

DELETED - Indicates the request to delete the environment is complete, and the environment has been deleted.

UNAVAILABLE - Indicates the request failed, but the environment was unable to rollback and is not in a stable state.

UPDATE_FAILED - Indicates the request to update the environment failed, and the environment has rolled back successfully and is ready to use.

We recommend reviewing our troubleshooting guide for a list of common errors and their solutions. To learn more, see Amazon MWAA troubleshooting.

minWorkers :: Maybe Natural

The minimum number of workers that run in your environment. For example, 2.

pluginsS3Path :: Maybe Text

The relative path to the plugins.zip file on your Amazon S3 bucket. For example, plugins.zip. To learn more, see Installing custom plugins.

webserverAccessMode :: Maybe WebserverAccessMode

The Apache Airflow Web server access mode. To learn more, see Apache Airflow access modes.

airflowVersion :: Maybe Text

The Apache Airflow version on your environment. For example, v1.10.12.

kmsKey :: Maybe Text

The Key Management Service (KMS) encryption key used to encrypt the data in your environment.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon MWAA environment.

createdAt :: Maybe POSIX

The day and time the environment was created.

weeklyMaintenanceWindowStart :: Maybe Text

The day and time of the week that weekly maintenance updates are scheduled. For example: TUE:03:30.

executionRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the execution role in IAM that allows MWAA to access AWS resources in your environment. For example, arn:aws:iam::123456789:role/my-execution-role. To learn more, see Amazon MWAA Execution role.

requirementsS3ObjectVersion :: Maybe Text

The version of the requirements.txt file on your Amazon S3 bucket. To learn more, see Installing Python dependencies.

lastUpdate :: Maybe LastUpdate

sourceBucketArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon S3 bucket where your DAG code and supporting files are stored. For example, arn:aws:s3:::my-airflow-bucket-unique-name. To learn more, see Create an Amazon S3 bucket for Amazon MWAA.

webserverUrl :: Maybe Text

The Apache Airflow Web server host name for the Amazon MWAA environment. To learn more, see Accessing the Apache Airflow UI.

dagS3Path :: Maybe Text

The relative path to the DAGs folder on your Amazon S3 bucket. For example, dags. To learn more, see Adding or updating DAGs.

name :: Maybe Text

The name of the Amazon MWAA environment. For example, MyMWAAEnvironment.

pluginsS3ObjectVersion :: Maybe Text

The version of the plugins.zip file on your Amazon S3 bucket. To learn more, see Installing custom plugins.

airflowConfigurationOptions :: Maybe (HashMap Text Text)

A list of key-value pairs containing the Apache Airflow configuration options attached to your environment. To learn more, see Apache Airflow configuration options.

loggingConfiguration :: Maybe LoggingConfiguration

The Apache Airflow logs being sent to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

environmentClass :: Maybe Text

The environment class type. Valid values: mw1.small, mw1.medium, mw1.large. To learn more, see Amazon MWAA environment class.

networkConfiguration :: Maybe NetworkConfiguration

tags :: Maybe (HashMap Text Text)

The key-value tag pairs associated to your environment. For example, "Environment": "Staging". To learn more, see Tagging AWS resources.

requirementsS3Path :: Maybe Text

The relative path to the requirements.txt file on your Amazon S3 bucket. For example, requirements.txt. To learn more, see Installing Python dependencies.

maxWorkers :: Maybe Natural

The maximum number of workers that run in your environment. For example, 20.

newEnvironment :: Environment

Create a value of Environment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceRoleArn:Environment', environment_serviceRoleArn - The Amazon Resource Name (ARN) for the service-linked role of the environment. To learn more, see Amazon MWAA Service-linked role.

$sel:schedulers:Environment', environment_schedulers - The number of Apache Airflow schedulers that run in your Amazon MWAA environment.

$sel:status:Environment', environment_status - The status of the Amazon MWAA environment. Valid values:

CREATING - Indicates the request to create the environment is in progress.

CREATE_FAILED - Indicates the request to create the environment failed, and the environment could not be created.

AVAILABLE - Indicates the request was successful and the environment is ready to use.

UPDATING - Indicates the request to update the environment is in progress.

DELETING - Indicates the request to delete the environment is in progress.

DELETED - Indicates the request to delete the environment is complete, and the environment has been deleted.

UNAVAILABLE - Indicates the request failed, but the environment was unable to rollback and is not in a stable state.

UPDATE_FAILED - Indicates the request to update the environment failed, and the environment has rolled back successfully and is ready to use.

We recommend reviewing our troubleshooting guide for a list of common errors and their solutions. To learn more, see Amazon MWAA troubleshooting.

$sel:minWorkers:Environment', environment_minWorkers - The minimum number of workers that run in your environment. For example, 2.

$sel:pluginsS3Path:Environment', environment_pluginsS3Path - The relative path to the plugins.zip file on your Amazon S3 bucket. For example, plugins.zip. To learn more, see Installing custom plugins.

$sel:webserverAccessMode:Environment', environment_webserverAccessMode - The Apache Airflow Web server access mode. To learn more, see Apache Airflow access modes.

$sel:airflowVersion:Environment', environment_airflowVersion - The Apache Airflow version on your environment. For example, v1.10.12.

$sel:kmsKey:Environment', environment_kmsKey - The Key Management Service (KMS) encryption key used to encrypt the data in your environment.

$sel:arn:Environment', environment_arn - The Amazon Resource Name (ARN) of the Amazon MWAA environment.

$sel:createdAt:Environment', environment_createdAt - The day and time the environment was created.

$sel:weeklyMaintenanceWindowStart:Environment', environment_weeklyMaintenanceWindowStart - The day and time of the week that weekly maintenance updates are scheduled. For example: TUE:03:30.

$sel:executionRoleArn:Environment', environment_executionRoleArn - The Amazon Resource Name (ARN) of the execution role in IAM that allows MWAA to access AWS resources in your environment. For example, arn:aws:iam::123456789:role/my-execution-role. To learn more, see Amazon MWAA Execution role.

$sel:requirementsS3ObjectVersion:Environment', environment_requirementsS3ObjectVersion - The version of the requirements.txt file on your Amazon S3 bucket. To learn more, see Installing Python dependencies.

$sel:lastUpdate:Environment', environment_lastUpdate - Undocumented member.

$sel:sourceBucketArn:Environment', environment_sourceBucketArn - The Amazon Resource Name (ARN) of the Amazon S3 bucket where your DAG code and supporting files are stored. For example, arn:aws:s3:::my-airflow-bucket-unique-name. To learn more, see Create an Amazon S3 bucket for Amazon MWAA.

$sel:webserverUrl:Environment', environment_webserverUrl - The Apache Airflow Web server host name for the Amazon MWAA environment. To learn more, see Accessing the Apache Airflow UI.

$sel:dagS3Path:Environment', environment_dagS3Path - The relative path to the DAGs folder on your Amazon S3 bucket. For example, dags. To learn more, see Adding or updating DAGs.

$sel:name:Environment', environment_name - The name of the Amazon MWAA environment. For example, MyMWAAEnvironment.

$sel:pluginsS3ObjectVersion:Environment', environment_pluginsS3ObjectVersion - The version of the plugins.zip file on your Amazon S3 bucket. To learn more, see Installing custom plugins.

$sel:airflowConfigurationOptions:Environment', environment_airflowConfigurationOptions - A list of key-value pairs containing the Apache Airflow configuration options attached to your environment. To learn more, see Apache Airflow configuration options.

$sel:loggingConfiguration:Environment', environment_loggingConfiguration - The Apache Airflow logs being sent to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

$sel:environmentClass:Environment', environment_environmentClass - The environment class type. Valid values: mw1.small, mw1.medium, mw1.large. To learn more, see Amazon MWAA environment class.

$sel:networkConfiguration:Environment', environment_networkConfiguration - Undocumented member.

$sel:tags:Environment', environment_tags - The key-value tag pairs associated to your environment. For example, "Environment": "Staging". To learn more, see Tagging AWS resources.

$sel:requirementsS3Path:Environment', environment_requirementsS3Path - The relative path to the requirements.txt file on your Amazon S3 bucket. For example, requirements.txt. To learn more, see Installing Python dependencies.

$sel:maxWorkers:Environment', environment_maxWorkers - The maximum number of workers that run in your environment. For example, 20.

environment_serviceRoleArn :: Lens' Environment (Maybe Text)

The Amazon Resource Name (ARN) for the service-linked role of the environment. To learn more, see Amazon MWAA Service-linked role.

environment_schedulers :: Lens' Environment (Maybe Int)

The number of Apache Airflow schedulers that run in your Amazon MWAA environment.

environment_status :: Lens' Environment (Maybe EnvironmentStatus)

The status of the Amazon MWAA environment. Valid values:

CREATING - Indicates the request to create the environment is in progress.

CREATE_FAILED - Indicates the request to create the environment failed, and the environment could not be created.

AVAILABLE - Indicates the request was successful and the environment is ready to use.

UPDATING - Indicates the request to update the environment is in progress.

DELETING - Indicates the request to delete the environment is in progress.

DELETED - Indicates the request to delete the environment is complete, and the environment has been deleted.

UNAVAILABLE - Indicates the request failed, but the environment was unable to rollback and is not in a stable state.

UPDATE_FAILED - Indicates the request to update the environment failed, and the environment has rolled back successfully and is ready to use.

We recommend reviewing our troubleshooting guide for a list of common errors and their solutions. To learn more, see Amazon MWAA troubleshooting.

environment_minWorkers :: Lens' Environment (Maybe Natural)

The minimum number of workers that run in your environment. For example, 2.

environment_pluginsS3Path :: Lens' Environment (Maybe Text)

The relative path to the plugins.zip file on your Amazon S3 bucket. For example, plugins.zip. To learn more, see Installing custom plugins.

environment_webserverAccessMode :: Lens' Environment (Maybe WebserverAccessMode)

The Apache Airflow Web server access mode. To learn more, see Apache Airflow access modes.

environment_airflowVersion :: Lens' Environment (Maybe Text)

The Apache Airflow version on your environment. For example, v1.10.12.

environment_kmsKey :: Lens' Environment (Maybe Text)

The Key Management Service (KMS) encryption key used to encrypt the data in your environment.

environment_arn :: Lens' Environment (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon MWAA environment.

environment_createdAt :: Lens' Environment (Maybe UTCTime)

The day and time the environment was created.

environment_weeklyMaintenanceWindowStart :: Lens' Environment (Maybe Text)

The day and time of the week that weekly maintenance updates are scheduled. For example: TUE:03:30.

environment_executionRoleArn :: Lens' Environment (Maybe Text)

The Amazon Resource Name (ARN) of the execution role in IAM that allows MWAA to access AWS resources in your environment. For example, arn:aws:iam::123456789:role/my-execution-role. To learn more, see Amazon MWAA Execution role.

environment_requirementsS3ObjectVersion :: Lens' Environment (Maybe Text)

The version of the requirements.txt file on your Amazon S3 bucket. To learn more, see Installing Python dependencies.

environment_lastUpdate :: Lens' Environment (Maybe LastUpdate)

Undocumented member.

environment_sourceBucketArn :: Lens' Environment (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon S3 bucket where your DAG code and supporting files are stored. For example, arn:aws:s3:::my-airflow-bucket-unique-name. To learn more, see Create an Amazon S3 bucket for Amazon MWAA.

environment_webserverUrl :: Lens' Environment (Maybe Text)

The Apache Airflow Web server host name for the Amazon MWAA environment. To learn more, see Accessing the Apache Airflow UI.

environment_dagS3Path :: Lens' Environment (Maybe Text)

The relative path to the DAGs folder on your Amazon S3 bucket. For example, dags. To learn more, see Adding or updating DAGs.

environment_name :: Lens' Environment (Maybe Text)

The name of the Amazon MWAA environment. For example, MyMWAAEnvironment.

environment_pluginsS3ObjectVersion :: Lens' Environment (Maybe Text)

The version of the plugins.zip file on your Amazon S3 bucket. To learn more, see Installing custom plugins.

environment_airflowConfigurationOptions :: Lens' Environment (Maybe (HashMap Text Text))

A list of key-value pairs containing the Apache Airflow configuration options attached to your environment. To learn more, see Apache Airflow configuration options.

environment_loggingConfiguration :: Lens' Environment (Maybe LoggingConfiguration)

The Apache Airflow logs being sent to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

environment_environmentClass :: Lens' Environment (Maybe Text)

The environment class type. Valid values: mw1.small, mw1.medium, mw1.large. To learn more, see Amazon MWAA environment class.

environment_networkConfiguration :: Lens' Environment (Maybe NetworkConfiguration)

Undocumented member.

environment_tags :: Lens' Environment (Maybe (HashMap Text Text))

The key-value tag pairs associated to your environment. For example, "Environment": "Staging". To learn more, see Tagging AWS resources.

environment_requirementsS3Path :: Lens' Environment (Maybe Text)

The relative path to the requirements.txt file on your Amazon S3 bucket. For example, requirements.txt. To learn more, see Installing Python dependencies.

environment_maxWorkers :: Lens' Environment (Maybe Natural)

The maximum number of workers that run in your environment. For example, 20.