Network.AWS.MwAA.Types.LoggingConfiguration

Description

Documentation

data LoggingConfiguration

Defines the Apache Airflow logs to send to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

See: newLoggingConfiguration smart constructor.

Constructors

LoggingConfiguration'

Fields

taskLogs :: Maybe ModuleLoggingConfiguration

webserverLogs :: Maybe ModuleLoggingConfiguration

schedulerLogs :: Maybe ModuleLoggingConfiguration

dagProcessingLogs :: Maybe ModuleLoggingConfiguration

workerLogs :: Maybe ModuleLoggingConfiguration

newLoggingConfiguration :: LoggingConfiguration

Create a value of LoggingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskLogs:LoggingConfiguration', loggingConfiguration_taskLogs - Undocumented member.

$sel:webserverLogs:LoggingConfiguration', loggingConfiguration_webserverLogs - Undocumented member.

$sel:schedulerLogs:LoggingConfiguration', loggingConfiguration_schedulerLogs - Undocumented member.

$sel:dagProcessingLogs:LoggingConfiguration', loggingConfiguration_dagProcessingLogs - Undocumented member.

$sel:workerLogs:LoggingConfiguration', loggingConfiguration_workerLogs - Undocumented member.

loggingConfiguration_taskLogs :: Lens' LoggingConfiguration (Maybe ModuleLoggingConfiguration)

Undocumented member.

loggingConfiguration_webserverLogs :: Lens' LoggingConfiguration (Maybe ModuleLoggingConfiguration)

Undocumented member.

loggingConfiguration_schedulerLogs :: Lens' LoggingConfiguration (Maybe ModuleLoggingConfiguration)

Undocumented member.

loggingConfiguration_dagProcessingLogs :: Lens' LoggingConfiguration (Maybe ModuleLoggingConfiguration)

Undocumented member.

loggingConfiguration_workerLogs :: Lens' LoggingConfiguration (Maybe ModuleLoggingConfiguration)

Undocumented member.