Network.AWS.MwAA.UpdateEnvironment

Description

Updates an Amazon Managed Workflows for Apache Airflow (MWAA) environment.

Creating a Request

data UpdateEnvironment

See: newUpdateEnvironment smart constructor.

Constructors

UpdateEnvironment'

Fields

schedulers :: Maybe Int

The number of Apache Airflow schedulers to run in your Amazon MWAA environment.

minWorkers :: Maybe Natural

The minimum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the worker count you specify in the MinWorkers field. For example, 2.

pluginsS3Path :: Maybe Text

The relative path to the plugins.zip file on your Amazon S3 bucket. For example, plugins.zip. If specified, then the plugins.zip version is required. To learn more, see Installing custom plugins.

webserverAccessMode :: Maybe WebserverAccessMode

The Apache Airflow Web server access mode. To learn more, see Apache Airflow access modes.

airflowVersion :: Maybe Text

The Apache Airflow version for your environment. For example, v1.10.12. If no value is specified, defaults to the latest version. Valid values: v1.10.12.

weeklyMaintenanceWindowStart :: Maybe Text

The day and time of the week to start weekly maintenance updates of your environment in the following format: DAY:HH:MM. For example: TUE:03:30. You can specify a start time in 30 minute increments only. Supported input includes the following:

MON|TUE|WED|THU|FRI|SAT|SUN:([01]\\d|2[0-3]):(00|30)

executionRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the execution role in IAM that allows MWAA to access AWS resources in your environment. For example, arn:aws:iam::123456789:role/my-execution-role. To learn more, see Amazon MWAA Execution role.

requirementsS3ObjectVersion :: Maybe Text

The version of the requirements.txt file on your Amazon S3 bucket. A version must be specified each time a requirements.txt file is updated. To learn more, see How S3 Versioning works.

sourceBucketArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon S3 bucket where your DAG code and supporting files are stored. For example, arn:aws:s3:::my-airflow-bucket-unique-name. To learn more, see Create an Amazon S3 bucket for Amazon MWAA.

dagS3Path :: Maybe Text

The relative path to the DAGs folder on your Amazon S3 bucket. For example, dags. To learn more, see Adding or updating DAGs.

pluginsS3ObjectVersion :: Maybe Text

The version of the plugins.zip file on your Amazon S3 bucket. A version must be specified each time a plugins.zip file is updated. To learn more, see How S3 Versioning works.

airflowConfigurationOptions :: Maybe (Sensitive (HashMap Text Text))

A list of key-value pairs containing the Apache Airflow configuration options you want to attach to your environment. To learn more, see Apache Airflow configuration options.

loggingConfiguration :: Maybe LoggingConfigurationInput

Defines the Apache Airflow logs to send to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

environmentClass :: Maybe Text

The environment class type. Valid values: mw1.small, mw1.medium, mw1.large. To learn more, see Amazon MWAA environment class.

networkConfiguration :: Maybe UpdateNetworkConfigurationInput

The VPC networking components used to secure and enable network traffic between the AWS resources for your environment. To learn more, see About networking on Amazon MWAA.

requirementsS3Path :: Maybe Text

The relative path to the requirements.txt file on your Amazon S3 bucket. For example, requirements.txt. If specified, then a file version is required. To learn more, see Installing Python dependencies.

maxWorkers :: Maybe Natural

The maximum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. For example, 20. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the one worker that is included with your environment, or the number you specify in MinWorkers.

name :: Text

The name of your Amazon MWAA environment. For example, MyMWAAEnvironment.

newUpdateEnvironment

Arguments

:: Text
$sel:name:UpdateEnvironment'

-> UpdateEnvironment

Create a value of UpdateEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schedulers:UpdateEnvironment', updateEnvironment_schedulers - The number of Apache Airflow schedulers to run in your Amazon MWAA environment.

$sel:minWorkers:UpdateEnvironment', updateEnvironment_minWorkers - The minimum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the worker count you specify in the MinWorkers field. For example, 2.

$sel:pluginsS3Path:UpdateEnvironment', updateEnvironment_pluginsS3Path - The relative path to the plugins.zip file on your Amazon S3 bucket. For example, plugins.zip. If specified, then the plugins.zip version is required. To learn more, see Installing custom plugins.

$sel:webserverAccessMode:UpdateEnvironment', updateEnvironment_webserverAccessMode - The Apache Airflow Web server access mode. To learn more, see Apache Airflow access modes.

$sel:airflowVersion:UpdateEnvironment', updateEnvironment_airflowVersion - The Apache Airflow version for your environment. For example, v1.10.12. If no value is specified, defaults to the latest version. Valid values: v1.10.12.

$sel:weeklyMaintenanceWindowStart:UpdateEnvironment', updateEnvironment_weeklyMaintenanceWindowStart - The day and time of the week to start weekly maintenance updates of your environment in the following format: DAY:HH:MM. For example: TUE:03:30. You can specify a start time in 30 minute increments only. Supported input includes the following:

MON|TUE|WED|THU|FRI|SAT|SUN:([01]\\d|2[0-3]):(00|30)

$sel:executionRoleArn:UpdateEnvironment', updateEnvironment_executionRoleArn - The Amazon Resource Name (ARN) of the execution role in IAM that allows MWAA to access AWS resources in your environment. For example, arn:aws:iam::123456789:role/my-execution-role. To learn more, see Amazon MWAA Execution role.

$sel:requirementsS3ObjectVersion:UpdateEnvironment', updateEnvironment_requirementsS3ObjectVersion - The version of the requirements.txt file on your Amazon S3 bucket. A version must be specified each time a requirements.txt file is updated. To learn more, see How S3 Versioning works.

$sel:sourceBucketArn:UpdateEnvironment', updateEnvironment_sourceBucketArn - The Amazon Resource Name (ARN) of the Amazon S3 bucket where your DAG code and supporting files are stored. For example, arn:aws:s3:::my-airflow-bucket-unique-name. To learn more, see Create an Amazon S3 bucket for Amazon MWAA.

$sel:dagS3Path:UpdateEnvironment', updateEnvironment_dagS3Path - The relative path to the DAGs folder on your Amazon S3 bucket. For example, dags. To learn more, see Adding or updating DAGs.

$sel:pluginsS3ObjectVersion:UpdateEnvironment', updateEnvironment_pluginsS3ObjectVersion - The version of the plugins.zip file on your Amazon S3 bucket. A version must be specified each time a plugins.zip file is updated. To learn more, see How S3 Versioning works.

$sel:airflowConfigurationOptions:UpdateEnvironment', updateEnvironment_airflowConfigurationOptions - A list of key-value pairs containing the Apache Airflow configuration options you want to attach to your environment. To learn more, see Apache Airflow configuration options.

$sel:loggingConfiguration:UpdateEnvironment', updateEnvironment_loggingConfiguration - Defines the Apache Airflow logs to send to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

$sel:environmentClass:UpdateEnvironment', updateEnvironment_environmentClass - The environment class type. Valid values: mw1.small, mw1.medium, mw1.large. To learn more, see Amazon MWAA environment class.

$sel:networkConfiguration:UpdateEnvironment', updateEnvironment_networkConfiguration - The VPC networking components used to secure and enable network traffic between the AWS resources for your environment. To learn more, see About networking on Amazon MWAA.

$sel:requirementsS3Path:UpdateEnvironment', updateEnvironment_requirementsS3Path - The relative path to the requirements.txt file on your Amazon S3 bucket. For example, requirements.txt. If specified, then a file version is required. To learn more, see Installing Python dependencies.

$sel:maxWorkers:UpdateEnvironment', updateEnvironment_maxWorkers - The maximum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. For example, 20. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the one worker that is included with your environment, or the number you specify in MinWorkers.

$sel:name:UpdateEnvironment', updateEnvironment_name - The name of your Amazon MWAA environment. For example, MyMWAAEnvironment.

Request Lenses

updateEnvironment_schedulers :: Lens' UpdateEnvironment (Maybe Int)

The number of Apache Airflow schedulers to run in your Amazon MWAA environment.

updateEnvironment_minWorkers :: Lens' UpdateEnvironment (Maybe Natural)

The minimum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the worker count you specify in the MinWorkers field. For example, 2.

updateEnvironment_pluginsS3Path :: Lens' UpdateEnvironment (Maybe Text)

The relative path to the plugins.zip file on your Amazon S3 bucket. For example, plugins.zip. If specified, then the plugins.zip version is required. To learn more, see Installing custom plugins.

updateEnvironment_webserverAccessMode :: Lens' UpdateEnvironment (Maybe WebserverAccessMode)

The Apache Airflow Web server access mode. To learn more, see Apache Airflow access modes.

updateEnvironment_airflowVersion :: Lens' UpdateEnvironment (Maybe Text)

The Apache Airflow version for your environment. For example, v1.10.12. If no value is specified, defaults to the latest version. Valid values: v1.10.12.

updateEnvironment_weeklyMaintenanceWindowStart :: Lens' UpdateEnvironment (Maybe Text)

The day and time of the week to start weekly maintenance updates of your environment in the following format: DAY:HH:MM. For example: TUE:03:30. You can specify a start time in 30 minute increments only. Supported input includes the following:

MON|TUE|WED|THU|FRI|SAT|SUN:([01]\\d|2[0-3]):(00|30)

updateEnvironment_executionRoleArn :: Lens' UpdateEnvironment (Maybe Text)

The Amazon Resource Name (ARN) of the execution role in IAM that allows MWAA to access AWS resources in your environment. For example, arn:aws:iam::123456789:role/my-execution-role. To learn more, see Amazon MWAA Execution role.

updateEnvironment_requirementsS3ObjectVersion :: Lens' UpdateEnvironment (Maybe Text)

The version of the requirements.txt file on your Amazon S3 bucket. A version must be specified each time a requirements.txt file is updated. To learn more, see How S3 Versioning works.

updateEnvironment_sourceBucketArn :: Lens' UpdateEnvironment (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon S3 bucket where your DAG code and supporting files are stored. For example, arn:aws:s3:::my-airflow-bucket-unique-name. To learn more, see Create an Amazon S3 bucket for Amazon MWAA.

updateEnvironment_dagS3Path :: Lens' UpdateEnvironment (Maybe Text)

The relative path to the DAGs folder on your Amazon S3 bucket. For example, dags. To learn more, see Adding or updating DAGs.

updateEnvironment_pluginsS3ObjectVersion :: Lens' UpdateEnvironment (Maybe Text)

The version of the plugins.zip file on your Amazon S3 bucket. A version must be specified each time a plugins.zip file is updated. To learn more, see How S3 Versioning works.

updateEnvironment_airflowConfigurationOptions :: Lens' UpdateEnvironment (Maybe (HashMap Text Text))

A list of key-value pairs containing the Apache Airflow configuration options you want to attach to your environment. To learn more, see Apache Airflow configuration options.

updateEnvironment_loggingConfiguration :: Lens' UpdateEnvironment (Maybe LoggingConfigurationInput)

Defines the Apache Airflow logs to send to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

updateEnvironment_environmentClass :: Lens' UpdateEnvironment (Maybe Text)

The environment class type. Valid values: mw1.small, mw1.medium, mw1.large. To learn more, see Amazon MWAA environment class.

updateEnvironment_networkConfiguration :: Lens' UpdateEnvironment (Maybe UpdateNetworkConfigurationInput)

The VPC networking components used to secure and enable network traffic between the AWS resources for your environment. To learn more, see About networking on Amazon MWAA.

updateEnvironment_requirementsS3Path :: Lens' UpdateEnvironment (Maybe Text)

The relative path to the requirements.txt file on your Amazon S3 bucket. For example, requirements.txt. If specified, then a file version is required. To learn more, see Installing Python dependencies.

updateEnvironment_maxWorkers :: Lens' UpdateEnvironment (Maybe Natural)

The maximum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. For example, 20. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the one worker that is included with your environment, or the number you specify in MinWorkers.

updateEnvironment_name :: Lens' UpdateEnvironment Text

The name of your Amazon MWAA environment. For example, MyMWAAEnvironment.

Destructuring the Response

data UpdateEnvironmentResponse

See: newUpdateEnvironmentResponse smart constructor.

Constructors

UpdateEnvironmentResponse'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon MWAA environment. For example, arn:aws:airflow:us-east-1:123456789012:environment/MyMWAAEnvironment.

httpStatus :: Int

The response's http status code.

newUpdateEnvironmentResponse

Arguments

:: Int
$sel:httpStatus:UpdateEnvironmentResponse'

-> UpdateEnvironmentResponse

Create a value of UpdateEnvironmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:UpdateEnvironmentResponse', updateEnvironmentResponse_arn - The Amazon Resource Name (ARN) of the Amazon MWAA environment. For example, arn:aws:airflow:us-east-1:123456789012:environment/MyMWAAEnvironment.

$sel:httpStatus:UpdateEnvironmentResponse', updateEnvironmentResponse_httpStatus - The response's http status code.

Response Lenses

updateEnvironmentResponse_arn :: Lens' UpdateEnvironmentResponse (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon MWAA environment. For example, arn:aws:airflow:us-east-1:123456789012:environment/MyMWAAEnvironment.

updateEnvironmentResponse_httpStatus :: Lens' UpdateEnvironmentResponse Int

The response's http status code.