Network.AWS.Neptune.CopyDBParameterGroup

Description

Copies the specified DB parameter group.

Creating a Request

data CopyDBParameterGroup

See: newCopyDBParameterGroup smart constructor.

Constructors

CopyDBParameterGroup'

Fields

tags :: Maybe [Tag]

The tags to be assigned to the copied DB parameter group.

sourceDBParameterGroupIdentifier :: Text

The identifier or ARN for the source DB parameter group. For information about creating an ARN, see Constructing an Amazon Resource Name (ARN).

Constraints:

Must specify a valid DB parameter group.

Must specify a valid DB parameter group identifier, for example my-db-param-group, or a valid ARN.

targetDBParameterGroupIdentifier :: Text

The identifier for the copied DB parameter group.

Constraints:

Cannot be null, empty, or blank.

Must contain from 1 to 255 letters, numbers, or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: my-db-parameter-group

targetDBParameterGroupDescription :: Text

A description for the copied DB parameter group.

newCopyDBParameterGroup

Arguments

:: Text
$sel:sourceDBParameterGroupIdentifier:CopyDBParameterGroup'

-> Text
$sel:targetDBParameterGroupIdentifier:CopyDBParameterGroup'

-> Text
$sel:targetDBParameterGroupDescription:CopyDBParameterGroup'

-> CopyDBParameterGroup

Create a value of CopyDBParameterGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CopyDBParameterGroup', copyDBParameterGroup_tags - The tags to be assigned to the copied DB parameter group.

$sel:sourceDBParameterGroupIdentifier:CopyDBParameterGroup', copyDBParameterGroup_sourceDBParameterGroupIdentifier - The identifier or ARN for the source DB parameter group. For information about creating an ARN, see Constructing an Amazon Resource Name (ARN).

Constraints:

Must specify a valid DB parameter group.

Must specify a valid DB parameter group identifier, for example my-db-param-group, or a valid ARN.

$sel:targetDBParameterGroupIdentifier:CopyDBParameterGroup', copyDBParameterGroup_targetDBParameterGroupIdentifier - The identifier for the copied DB parameter group.

Constraints:

Cannot be null, empty, or blank.

Must contain from 1 to 255 letters, numbers, or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: my-db-parameter-group

$sel:targetDBParameterGroupDescription:CopyDBParameterGroup', copyDBParameterGroup_targetDBParameterGroupDescription - A description for the copied DB parameter group.

Request Lenses

copyDBParameterGroup_tags :: Lens' CopyDBParameterGroup (Maybe [Tag])

The tags to be assigned to the copied DB parameter group.

copyDBParameterGroup_sourceDBParameterGroupIdentifier :: Lens' CopyDBParameterGroup Text

The identifier or ARN for the source DB parameter group. For information about creating an ARN, see Constructing an Amazon Resource Name (ARN).

Constraints:

Must specify a valid DB parameter group.

Must specify a valid DB parameter group identifier, for example my-db-param-group, or a valid ARN.

copyDBParameterGroup_targetDBParameterGroupIdentifier :: Lens' CopyDBParameterGroup Text

The identifier for the copied DB parameter group.

Constraints:

Cannot be null, empty, or blank.

Must contain from 1 to 255 letters, numbers, or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: my-db-parameter-group

copyDBParameterGroup_targetDBParameterGroupDescription :: Lens' CopyDBParameterGroup Text

A description for the copied DB parameter group.

Destructuring the Response

data CopyDBParameterGroupResponse

See: newCopyDBParameterGroupResponse smart constructor.

Constructors

CopyDBParameterGroupResponse'

Fields

dbParameterGroup :: Maybe DBParameterGroup

httpStatus :: Int

The response's http status code.

newCopyDBParameterGroupResponse

Arguments

:: Int
$sel:httpStatus:CopyDBParameterGroupResponse'

-> CopyDBParameterGroupResponse

Create a value of CopyDBParameterGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbParameterGroup:CopyDBParameterGroupResponse', copyDBParameterGroupResponse_dbParameterGroup - Undocumented member.

$sel:httpStatus:CopyDBParameterGroupResponse', copyDBParameterGroupResponse_httpStatus - The response's http status code.

Response Lenses

copyDBParameterGroupResponse_dbParameterGroup :: Lens' CopyDBParameterGroupResponse (Maybe DBParameterGroup)

Undocumented member.

copyDBParameterGroupResponse_httpStatus :: Lens' CopyDBParameterGroupResponse Int

The response's http status code.