Network.AWS.Neptune.CreateDBCluster

Description

Creates a new Amazon Neptune DB cluster.

You can use the ReplicationSourceIdentifier parameter to create the DB cluster as a Read Replica of another DB cluster or Amazon Neptune DB instance.

Note that when you create a new cluster using CreateDBCluster directly, deletion protection is disabled by default (when you create a new production cluster in the console, deletion protection is enabled by default). You can only delete a DB cluster if its DeletionProtection field is set to false.

Creating a Request

data CreateDBCluster

See: newCreateDBCluster smart constructor.

Constructors

CreateDBCluster'

Fields

engineVersion :: Maybe Text

The version number of the database engine to use for the new DB cluster.

Example: 1.0.2.1

deletionProtection :: Maybe Bool

A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is enabled.

storageEncrypted :: Maybe Bool

Specifies whether the DB cluster is encrypted.

masterUserPassword :: Maybe Text

Not supported by Neptune.

replicationSourceIdentifier :: Maybe Text

The Amazon Resource Name (ARN) of the source DB instance or DB cluster if this DB cluster is created as a Read Replica.

masterUsername :: Maybe Text

Not supported by Neptune.

dbSubnetGroupName :: Maybe Text

A DB subnet group to associate with this DB cluster.

Constraints: Must match the name of an existing DBSubnetGroup. Must not be default.

Example: mySubnetgroup

preSignedUrl :: Maybe Text

This parameter is not currently supported.

preferredMaintenanceWindow :: Maybe Text

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon Neptune User Guide.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

availabilityZones :: Maybe [Text]

A list of EC2 Availability Zones that instances in the DB cluster can be created in.

characterSetName :: Maybe Text

(Not supported by Neptune)

kmsKeyId :: Maybe Text

The Amazon KMS key identifier for an encrypted DB cluster.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a DB cluster with the same Amazon account that owns the KMS encryption key used to encrypt the new DB cluster, then you can use the KMS key alias instead of the ARN for the KMS encryption key.

If an encryption key is not specified in KmsKeyId:

If ReplicationSourceIdentifier identifies an encrypted source, then Amazon Neptune will use the encryption key used to encrypt the source. Otherwise, Amazon Neptune will use your default encryption key.

If the StorageEncrypted parameter is true and ReplicationSourceIdentifier is not specified, then Amazon Neptune will use your default encryption key.

Amazon KMS creates the default encryption key for your Amazon account. Your Amazon account has a different default encryption key for each Amazon Region.

If you create a Read Replica of an encrypted DB cluster in another Amazon Region, you must set KmsKeyId to a KMS key ID that is valid in the destination Amazon Region. This key is used to encrypt the Read Replica in that Amazon Region.

preferredBackupWindow :: Maybe Text

The daily time range during which automated backups are created if automated backups are enabled using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Region. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon Neptune User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

backupRetentionPeriod :: Maybe Int

The number of days for which automated backups are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35

vpcSecurityGroupIds :: Maybe [Text]

A list of EC2 VPC security groups to associate with this DB cluster.

databaseName :: Maybe Text

The name for your database of up to 64 alpha-numeric characters. If you do not provide a name, Amazon Neptune will not create a database in the DB cluster you are creating.

dbClusterParameterGroupName :: Maybe Text

The name of the DB cluster parameter group to associate with this DB cluster. If this argument is omitted, the default is used.

Constraints:

If supplied, must match the name of an existing DBClusterParameterGroup.

optionGroupName :: Maybe Text

(Not supported by Neptune)

copyTagsToSnapshot :: Maybe Bool

/If set to true, tags are copied to any snapshot of the DB cluster that is created./

tags :: Maybe [Tag]

The tags to assign to the new DB cluster.

port :: Maybe Int

The port number on which the instances in the DB cluster accept connections.

Default: 8182

enableIAMDatabaseAuthentication :: Maybe Bool

If set to true, enables Amazon Identity and Access Management (IAM) authentication for the entire DB cluster (this cannot be set at an instance level).

Default: false.

enableCloudwatchLogsExports :: Maybe [Text]

The list of log types that need to be enabled for exporting to CloudWatch Logs.

dbClusterIdentifier :: Text

The DB cluster identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: my-cluster1

engine :: Text

The name of the database engine to be used for this DB cluster.

Valid Values: neptune

newCreateDBCluster

Arguments

:: Text
$sel:dbClusterIdentifier:CreateDBCluster'

-> Text
$sel:engine:CreateDBCluster'

-> CreateDBCluster

Create a value of CreateDBCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:CreateDBCluster', createDBCluster_engineVersion - The version number of the database engine to use for the new DB cluster.

Example: 1.0.2.1

$sel:deletionProtection:CreateDBCluster', createDBCluster_deletionProtection - A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is enabled.

$sel:storageEncrypted:CreateDBCluster', createDBCluster_storageEncrypted - Specifies whether the DB cluster is encrypted.

$sel:masterUserPassword:CreateDBCluster', createDBCluster_masterUserPassword - Not supported by Neptune.

$sel:replicationSourceIdentifier:CreateDBCluster', createDBCluster_replicationSourceIdentifier - The Amazon Resource Name (ARN) of the source DB instance or DB cluster if this DB cluster is created as a Read Replica.

$sel:masterUsername:CreateDBCluster', createDBCluster_masterUsername - Not supported by Neptune.

$sel:dbSubnetGroupName:CreateDBCluster', createDBCluster_dbSubnetGroupName - A DB subnet group to associate with this DB cluster.

Constraints: Must match the name of an existing DBSubnetGroup. Must not be default.

Example: mySubnetgroup

$sel:preSignedUrl:CreateDBCluster', createDBCluster_preSignedUrl - This parameter is not currently supported.

$sel:preferredMaintenanceWindow:CreateDBCluster', createDBCluster_preferredMaintenanceWindow - The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon Neptune User Guide.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

$sel:availabilityZones:CreateDBCluster', createDBCluster_availabilityZones - A list of EC2 Availability Zones that instances in the DB cluster can be created in.

$sel:characterSetName:CreateDBCluster', createDBCluster_characterSetName - (Not supported by Neptune)

$sel:kmsKeyId:CreateDBCluster', createDBCluster_kmsKeyId - The Amazon KMS key identifier for an encrypted DB cluster.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a DB cluster with the same Amazon account that owns the KMS encryption key used to encrypt the new DB cluster, then you can use the KMS key alias instead of the ARN for the KMS encryption key.

If an encryption key is not specified in KmsKeyId:

If ReplicationSourceIdentifier identifies an encrypted source, then Amazon Neptune will use the encryption key used to encrypt the source. Otherwise, Amazon Neptune will use your default encryption key.

If the StorageEncrypted parameter is true and ReplicationSourceIdentifier is not specified, then Amazon Neptune will use your default encryption key.

Amazon KMS creates the default encryption key for your Amazon account. Your Amazon account has a different default encryption key for each Amazon Region.

If you create a Read Replica of an encrypted DB cluster in another Amazon Region, you must set KmsKeyId to a KMS key ID that is valid in the destination Amazon Region. This key is used to encrypt the Read Replica in that Amazon Region.

$sel:preferredBackupWindow:CreateDBCluster', createDBCluster_preferredBackupWindow - The daily time range during which automated backups are created if automated backups are enabled using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Region. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon Neptune User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

$sel:backupRetentionPeriod:CreateDBCluster', createDBCluster_backupRetentionPeriod - The number of days for which automated backups are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35

$sel:vpcSecurityGroupIds:CreateDBCluster', createDBCluster_vpcSecurityGroupIds - A list of EC2 VPC security groups to associate with this DB cluster.

$sel:databaseName:CreateDBCluster', createDBCluster_databaseName - The name for your database of up to 64 alpha-numeric characters. If you do not provide a name, Amazon Neptune will not create a database in the DB cluster you are creating.

$sel:dbClusterParameterGroupName:CreateDBCluster', createDBCluster_dbClusterParameterGroupName - The name of the DB cluster parameter group to associate with this DB cluster. If this argument is omitted, the default is used.

Constraints:

If supplied, must match the name of an existing DBClusterParameterGroup.

$sel:optionGroupName:CreateDBCluster', createDBCluster_optionGroupName - (Not supported by Neptune)

$sel:copyTagsToSnapshot:CreateDBCluster', createDBCluster_copyTagsToSnapshot - /If set to true, tags are copied to any snapshot of the DB cluster that is created./

$sel:tags:CreateDBCluster', createDBCluster_tags - The tags to assign to the new DB cluster.

$sel:port:CreateDBCluster', createDBCluster_port - The port number on which the instances in the DB cluster accept connections.

Default: 8182

$sel:enableIAMDatabaseAuthentication:CreateDBCluster', createDBCluster_enableIAMDatabaseAuthentication - If set to true, enables Amazon Identity and Access Management (IAM) authentication for the entire DB cluster (this cannot be set at an instance level).

Default: false.

$sel:enableCloudwatchLogsExports:CreateDBCluster', createDBCluster_enableCloudwatchLogsExports - The list of log types that need to be enabled for exporting to CloudWatch Logs.

$sel:dbClusterIdentifier:CreateDBCluster', createDBCluster_dbClusterIdentifier - The DB cluster identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: my-cluster1

$sel:engine:CreateDBCluster', createDBCluster_engine - The name of the database engine to be used for this DB cluster.

Valid Values: neptune

Request Lenses

createDBCluster_engineVersion :: Lens' CreateDBCluster (Maybe Text)

The version number of the database engine to use for the new DB cluster.

Example: 1.0.2.1

createDBCluster_deletionProtection :: Lens' CreateDBCluster (Maybe Bool)

A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is enabled.

createDBCluster_storageEncrypted :: Lens' CreateDBCluster (Maybe Bool)

Specifies whether the DB cluster is encrypted.

createDBCluster_masterUserPassword :: Lens' CreateDBCluster (Maybe Text)

Not supported by Neptune.

createDBCluster_replicationSourceIdentifier :: Lens' CreateDBCluster (Maybe Text)

The Amazon Resource Name (ARN) of the source DB instance or DB cluster if this DB cluster is created as a Read Replica.

createDBCluster_masterUsername :: Lens' CreateDBCluster (Maybe Text)

Not supported by Neptune.

createDBCluster_dbSubnetGroupName :: Lens' CreateDBCluster (Maybe Text)

A DB subnet group to associate with this DB cluster.

Constraints: Must match the name of an existing DBSubnetGroup. Must not be default.

Example: mySubnetgroup

createDBCluster_preSignedUrl :: Lens' CreateDBCluster (Maybe Text)

This parameter is not currently supported.

createDBCluster_preferredMaintenanceWindow :: Lens' CreateDBCluster (Maybe Text)

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon Neptune User Guide.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

createDBCluster_availabilityZones :: Lens' CreateDBCluster (Maybe [Text])

A list of EC2 Availability Zones that instances in the DB cluster can be created in.

createDBCluster_characterSetName :: Lens' CreateDBCluster (Maybe Text)

(Not supported by Neptune)

createDBCluster_kmsKeyId :: Lens' CreateDBCluster (Maybe Text)

The Amazon KMS key identifier for an encrypted DB cluster.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a DB cluster with the same Amazon account that owns the KMS encryption key used to encrypt the new DB cluster, then you can use the KMS key alias instead of the ARN for the KMS encryption key.

If an encryption key is not specified in KmsKeyId:

If ReplicationSourceIdentifier identifies an encrypted source, then Amazon Neptune will use the encryption key used to encrypt the source. Otherwise, Amazon Neptune will use your default encryption key.

If the StorageEncrypted parameter is true and ReplicationSourceIdentifier is not specified, then Amazon Neptune will use your default encryption key.

Amazon KMS creates the default encryption key for your Amazon account. Your Amazon account has a different default encryption key for each Amazon Region.

If you create a Read Replica of an encrypted DB cluster in another Amazon Region, you must set KmsKeyId to a KMS key ID that is valid in the destination Amazon Region. This key is used to encrypt the Read Replica in that Amazon Region.

createDBCluster_preferredBackupWindow :: Lens' CreateDBCluster (Maybe Text)

The daily time range during which automated backups are created if automated backups are enabled using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Region. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon Neptune User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

createDBCluster_backupRetentionPeriod :: Lens' CreateDBCluster (Maybe Int)

The number of days for which automated backups are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35

createDBCluster_vpcSecurityGroupIds :: Lens' CreateDBCluster (Maybe [Text])

A list of EC2 VPC security groups to associate with this DB cluster.

createDBCluster_databaseName :: Lens' CreateDBCluster (Maybe Text)

The name for your database of up to 64 alpha-numeric characters. If you do not provide a name, Amazon Neptune will not create a database in the DB cluster you are creating.

createDBCluster_dbClusterParameterGroupName :: Lens' CreateDBCluster (Maybe Text)

The name of the DB cluster parameter group to associate with this DB cluster. If this argument is omitted, the default is used.

Constraints:

If supplied, must match the name of an existing DBClusterParameterGroup.

createDBCluster_optionGroupName :: Lens' CreateDBCluster (Maybe Text)

(Not supported by Neptune)

createDBCluster_copyTagsToSnapshot :: Lens' CreateDBCluster (Maybe Bool)

/If set to true, tags are copied to any snapshot of the DB cluster that is created./

createDBCluster_tags :: Lens' CreateDBCluster (Maybe [Tag])

The tags to assign to the new DB cluster.

createDBCluster_port :: Lens' CreateDBCluster (Maybe Int)

The port number on which the instances in the DB cluster accept connections.

Default: 8182

createDBCluster_enableIAMDatabaseAuthentication :: Lens' CreateDBCluster (Maybe Bool)

If set to true, enables Amazon Identity and Access Management (IAM) authentication for the entire DB cluster (this cannot be set at an instance level).

Default: false.

createDBCluster_enableCloudwatchLogsExports :: Lens' CreateDBCluster (Maybe [Text])

The list of log types that need to be enabled for exporting to CloudWatch Logs.

createDBCluster_dbClusterIdentifier :: Lens' CreateDBCluster Text

The DB cluster identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: my-cluster1

createDBCluster_engine :: Lens' CreateDBCluster Text

The name of the database engine to be used for this DB cluster.

Valid Values: neptune

Destructuring the Response

data CreateDBClusterResponse

See: newCreateDBClusterResponse smart constructor.

Constructors

CreateDBClusterResponse'

Fields

dbCluster :: Maybe DBCluster

httpStatus :: Int

The response's http status code.

newCreateDBClusterResponse

Arguments

:: Int
$sel:httpStatus:CreateDBClusterResponse'

-> CreateDBClusterResponse

Create a value of CreateDBClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbCluster:CreateDBClusterResponse', createDBClusterResponse_dbCluster - Undocumented member.

$sel:httpStatus:CreateDBClusterResponse', createDBClusterResponse_httpStatus - The response's http status code.

Response Lenses

createDBClusterResponse_dbCluster :: Lens' CreateDBClusterResponse (Maybe DBCluster)

Undocumented member.

createDBClusterResponse_httpStatus :: Lens' CreateDBClusterResponse Int

The response's http status code.