Network.AWS.Neptune.CreateDBInstance

Description

Creates a new DB instance.

Creating a Request

data CreateDBInstance

See: newCreateDBInstance smart constructor.

Constructors

CreateDBInstance'

Fields

engineVersion :: Maybe Text

The version number of the database engine to use. Currently, setting this parameter has no effect.

dbSecurityGroups :: Maybe [Text]

A list of DB security groups to associate with this DB instance.

Default: The default DB security group for the database engine.

deletionProtection :: Maybe Bool

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. See Deleting a DB Instance.

DB instances in a DB cluster can be deleted even when deletion protection is enabled in their parent DB cluster.

storageEncrypted :: Maybe Bool

Specifies whether the DB instance is encrypted.

Not applicable. The encryption for DB instances is managed by the DB cluster. For more information, see CreateDBCluster.

Default: false

dbClusterIdentifier :: Maybe Text

The identifier of the DB cluster that the instance will belong to.

For information on creating a DB cluster, see CreateDBCluster.

Type: String

masterUserPassword :: Maybe Text

Not supported by Neptune.

publiclyAccessible :: Maybe Bool

This flag should no longer be used.

autoMinorVersionUpgrade :: Maybe Bool

Indicates that minor engine upgrades are applied automatically to the DB instance during the maintenance window.

Default: true

masterUsername :: Maybe Text

Not supported by Neptune.

dbSubnetGroupName :: Maybe Text

A DB subnet group to associate with this DB instance.

If there is no DB subnet group, then it is a non-VPC DB instance.

monitoringRoleArn :: Maybe Text

The ARN for the IAM role that permits Neptune to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

iops :: Maybe Int

The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance.

domain :: Maybe Text

Specify the Active Directory Domain to create the instance in.

monitoringInterval :: Maybe Int

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance. To disable collecting Enhanced Monitoring metrics, specify 0. The default is 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

tdeCredentialPassword :: Maybe Text

The password for the given ARN from the key store in order to access the device.

promotionTier :: Maybe Int

A value that specifies the order in which an Read Replica is promoted to the primary instance after a failure of the existing primary instance.

Default: 1

Valid Values: 0 - 15

licenseModel :: Maybe Text

License model information for this DB instance.

Valid values: license-included | bring-your-own-license | general-public-license

preferredMaintenanceWindow :: Maybe Text

The time range each week during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Region, occurring on a random day of the week.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

characterSetName :: Maybe Text

(Not supported by Neptune)

enablePerformanceInsights :: Maybe Bool

(Not supported by Neptune)

kmsKeyId :: Maybe Text

The Amazon KMS key identifier for an encrypted DB instance.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a DB instance with the same Amazon account that owns the KMS encryption key used to encrypt the new DB instance, then you can use the KMS key alias instead of the ARN for the KM encryption key.

Not applicable. The KMS key identifier is managed by the DB cluster. For more information, see CreateDBCluster.

If the StorageEncrypted parameter is true, and you do not specify a value for the KmsKeyId parameter, then Amazon Neptune will use your default encryption key. Amazon KMS creates the default encryption key for your Amazon account. Your Amazon account has a different default encryption key for each Amazon Region.

dbParameterGroupName :: Maybe Text

The name of the DB parameter group to associate with this DB instance. If this argument is omitted, the default DBParameterGroup for the specified engine is used.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

preferredBackupWindow :: Maybe Text

The daily time range during which automated backups are created.

Not applicable. The daily time range for creating automated backups is managed by the DB cluster. For more information, see CreateDBCluster.

availabilityZone :: Maybe Text

The EC2 Availability Zone that the DB instance is created in

Default: A random, system-chosen Availability Zone in the endpoint's Amazon Region.

Example: us-east-1d

Constraint: The AvailabilityZone parameter can't be specified if the MultiAZ parameter is set to true. The specified Availability Zone must be in the same Amazon Region as the current endpoint.

backupRetentionPeriod :: Maybe Int

The number of days for which automated backups are retained.

Not applicable. The retention period for automated backups is managed by the DB cluster. For more information, see CreateDBCluster.

Default: 1

Constraints:

Must be a value from 0 to 35

Cannot be set to 0 if the DB instance is a source to Read Replicas

performanceInsightsKMSKeyId :: Maybe Text

(Not supported by Neptune)

vpcSecurityGroupIds :: Maybe [Text]

A list of EC2 VPC security groups to associate with this DB instance.

Not applicable. The associated list of EC2 VPC security groups is managed by the DB cluster. For more information, see CreateDBCluster.

Default: The default EC2 VPC security group for the DB subnet group's VPC.

multiAZ :: Maybe Bool

Specifies if the DB instance is a Multi-AZ deployment. You can't set the AvailabilityZone parameter if the MultiAZ parameter is set to true.

allocatedStorage :: Maybe Int

Not supported by Neptune.

optionGroupName :: Maybe Text

(Not supported by Neptune)

copyTagsToSnapshot :: Maybe Bool

True to copy all tags from the DB instance to snapshots of the DB instance, and otherwise false. The default is false.

timezone :: Maybe Text

The time zone of the DB instance.

tdeCredentialArn :: Maybe Text

The ARN from the key store with which to associate the instance for TDE encryption.

domainIAMRoleName :: Maybe Text

Specify the name of the IAM role to be used when making API calls to the Directory Service.

tags :: Maybe [Tag]

The tags to assign to the new instance.

port :: Maybe Int

The port number on which the database accepts connections.

Not applicable. The port is managed by the DB cluster. For more information, see CreateDBCluster.

Default: 8182

Type: Integer

enableIAMDatabaseAuthentication :: Maybe Bool

Not supported by Neptune (ignored).

storageType :: Maybe Text

Specifies the storage type to be associated with the DB instance.

Not applicable. Storage is managed by the DB Cluster.

enableCloudwatchLogsExports :: Maybe [Text]

The list of log types that need to be enabled for exporting to CloudWatch Logs.

dbName :: Maybe Text

Not supported.

dbInstanceIdentifier :: Text

The DB instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

dbInstanceClass :: Text

The compute and memory capacity of the DB instance, for example, db.m4.large. Not all DB instance classes are available in all Amazon Regions.

engine :: Text

The name of the database engine to be used for this instance.

Valid Values: neptune

newCreateDBInstance

Arguments

:: Text
$sel:dbInstanceIdentifier:CreateDBInstance'

-> Text
$sel:dbInstanceClass:CreateDBInstance'

-> Text
$sel:engine:CreateDBInstance'

-> CreateDBInstance

Create a value of CreateDBInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:CreateDBInstance', createDBInstance_engineVersion - The version number of the database engine to use. Currently, setting this parameter has no effect.

$sel:dbSecurityGroups:CreateDBInstance', createDBInstance_dbSecurityGroups - A list of DB security groups to associate with this DB instance.

Default: The default DB security group for the database engine.

$sel:deletionProtection:CreateDBInstance', createDBInstance_deletionProtection - A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. See Deleting a DB Instance.

DB instances in a DB cluster can be deleted even when deletion protection is enabled in their parent DB cluster.

$sel:storageEncrypted:CreateDBInstance', createDBInstance_storageEncrypted - Specifies whether the DB instance is encrypted.

Not applicable. The encryption for DB instances is managed by the DB cluster. For more information, see CreateDBCluster.

Default: false

$sel:dbClusterIdentifier:CreateDBInstance', createDBInstance_dbClusterIdentifier - The identifier of the DB cluster that the instance will belong to.

For information on creating a DB cluster, see CreateDBCluster.

Type: String

$sel:masterUserPassword:CreateDBInstance', createDBInstance_masterUserPassword - Not supported by Neptune.

$sel:publiclyAccessible:CreateDBInstance', createDBInstance_publiclyAccessible - This flag should no longer be used.

$sel:autoMinorVersionUpgrade:CreateDBInstance', createDBInstance_autoMinorVersionUpgrade - Indicates that minor engine upgrades are applied automatically to the DB instance during the maintenance window.

Default: true

$sel:masterUsername:CreateDBInstance', createDBInstance_masterUsername - Not supported by Neptune.

$sel:dbSubnetGroupName:CreateDBInstance', createDBInstance_dbSubnetGroupName - A DB subnet group to associate with this DB instance.

If there is no DB subnet group, then it is a non-VPC DB instance.

$sel:monitoringRoleArn:CreateDBInstance', createDBInstance_monitoringRoleArn - The ARN for the IAM role that permits Neptune to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

$sel:iops:CreateDBInstance', createDBInstance_iops - The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance.

$sel:domain:CreateDBInstance', createDBInstance_domain - Specify the Active Directory Domain to create the instance in.

$sel:monitoringInterval:CreateDBInstance', createDBInstance_monitoringInterval - The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance. To disable collecting Enhanced Monitoring metrics, specify 0. The default is 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

$sel:tdeCredentialPassword:CreateDBInstance', createDBInstance_tdeCredentialPassword - The password for the given ARN from the key store in order to access the device.

$sel:promotionTier:CreateDBInstance', createDBInstance_promotionTier - A value that specifies the order in which an Read Replica is promoted to the primary instance after a failure of the existing primary instance.

Default: 1

Valid Values: 0 - 15

$sel:licenseModel:CreateDBInstance', createDBInstance_licenseModel - License model information for this DB instance.

Valid values: license-included | bring-your-own-license | general-public-license

$sel:preferredMaintenanceWindow:CreateDBInstance', createDBInstance_preferredMaintenanceWindow - The time range each week during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Region, occurring on a random day of the week.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

$sel:characterSetName:CreateDBInstance', createDBInstance_characterSetName - (Not supported by Neptune)

$sel:enablePerformanceInsights:CreateDBInstance', createDBInstance_enablePerformanceInsights - (Not supported by Neptune)

$sel:kmsKeyId:CreateDBInstance', createDBInstance_kmsKeyId - The Amazon KMS key identifier for an encrypted DB instance.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a DB instance with the same Amazon account that owns the KMS encryption key used to encrypt the new DB instance, then you can use the KMS key alias instead of the ARN for the KM encryption key.

Not applicable. The KMS key identifier is managed by the DB cluster. For more information, see CreateDBCluster.

If the StorageEncrypted parameter is true, and you do not specify a value for the KmsKeyId parameter, then Amazon Neptune will use your default encryption key. Amazon KMS creates the default encryption key for your Amazon account. Your Amazon account has a different default encryption key for each Amazon Region.

$sel:dbParameterGroupName:CreateDBInstance', createDBInstance_dbParameterGroupName - The name of the DB parameter group to associate with this DB instance. If this argument is omitted, the default DBParameterGroup for the specified engine is used.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

$sel:preferredBackupWindow:CreateDBInstance', createDBInstance_preferredBackupWindow - The daily time range during which automated backups are created.

Not applicable. The daily time range for creating automated backups is managed by the DB cluster. For more information, see CreateDBCluster.

$sel:availabilityZone:CreateDBInstance', createDBInstance_availabilityZone - The EC2 Availability Zone that the DB instance is created in

Default: A random, system-chosen Availability Zone in the endpoint's Amazon Region.

Example: us-east-1d

Constraint: The AvailabilityZone parameter can't be specified if the MultiAZ parameter is set to true. The specified Availability Zone must be in the same Amazon Region as the current endpoint.

$sel:backupRetentionPeriod:CreateDBInstance', createDBInstance_backupRetentionPeriod - The number of days for which automated backups are retained.

Not applicable. The retention period for automated backups is managed by the DB cluster. For more information, see CreateDBCluster.

Default: 1

Constraints:

Must be a value from 0 to 35

Cannot be set to 0 if the DB instance is a source to Read Replicas

$sel:performanceInsightsKMSKeyId:CreateDBInstance', createDBInstance_performanceInsightsKMSKeyId - (Not supported by Neptune)

$sel:vpcSecurityGroupIds:CreateDBInstance', createDBInstance_vpcSecurityGroupIds - A list of EC2 VPC security groups to associate with this DB instance.

Not applicable. The associated list of EC2 VPC security groups is managed by the DB cluster. For more information, see CreateDBCluster.

Default: The default EC2 VPC security group for the DB subnet group's VPC.

$sel:multiAZ:CreateDBInstance', createDBInstance_multiAZ - Specifies if the DB instance is a Multi-AZ deployment. You can't set the AvailabilityZone parameter if the MultiAZ parameter is set to true.

$sel:allocatedStorage:CreateDBInstance', createDBInstance_allocatedStorage - Not supported by Neptune.

$sel:optionGroupName:CreateDBInstance', createDBInstance_optionGroupName - (Not supported by Neptune)

$sel:copyTagsToSnapshot:CreateDBInstance', createDBInstance_copyTagsToSnapshot - True to copy all tags from the DB instance to snapshots of the DB instance, and otherwise false. The default is false.

$sel:timezone:CreateDBInstance', createDBInstance_timezone - The time zone of the DB instance.

$sel:tdeCredentialArn:CreateDBInstance', createDBInstance_tdeCredentialArn - The ARN from the key store with which to associate the instance for TDE encryption.

$sel:domainIAMRoleName:CreateDBInstance', createDBInstance_domainIAMRoleName - Specify the name of the IAM role to be used when making API calls to the Directory Service.

$sel:tags:CreateDBInstance', createDBInstance_tags - The tags to assign to the new instance.

$sel:port:CreateDBInstance', createDBInstance_port - The port number on which the database accepts connections.

Not applicable. The port is managed by the DB cluster. For more information, see CreateDBCluster.

Default: 8182

Type: Integer

$sel:enableIAMDatabaseAuthentication:CreateDBInstance', createDBInstance_enableIAMDatabaseAuthentication - Not supported by Neptune (ignored).

$sel:storageType:CreateDBInstance', createDBInstance_storageType - Specifies the storage type to be associated with the DB instance.

Not applicable. Storage is managed by the DB Cluster.

$sel:enableCloudwatchLogsExports:CreateDBInstance', createDBInstance_enableCloudwatchLogsExports - The list of log types that need to be enabled for exporting to CloudWatch Logs.

$sel:dbName:CreateDBInstance', createDBInstance_dbName - Not supported.

$sel:dbInstanceIdentifier:CreateDBInstance', createDBInstance_dbInstanceIdentifier - The DB instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

$sel:dbInstanceClass:CreateDBInstance', createDBInstance_dbInstanceClass - The compute and memory capacity of the DB instance, for example, db.m4.large. Not all DB instance classes are available in all Amazon Regions.

$sel:engine:CreateDBInstance', createDBInstance_engine - The name of the database engine to be used for this instance.

Valid Values: neptune

Request Lenses

createDBInstance_engineVersion :: Lens' CreateDBInstance (Maybe Text)

The version number of the database engine to use. Currently, setting this parameter has no effect.

createDBInstance_dbSecurityGroups :: Lens' CreateDBInstance (Maybe [Text])

A list of DB security groups to associate with this DB instance.

Default: The default DB security group for the database engine.

createDBInstance_deletionProtection :: Lens' CreateDBInstance (Maybe Bool)

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. See Deleting a DB Instance.

DB instances in a DB cluster can be deleted even when deletion protection is enabled in their parent DB cluster.

createDBInstance_storageEncrypted :: Lens' CreateDBInstance (Maybe Bool)

Specifies whether the DB instance is encrypted.

Not applicable. The encryption for DB instances is managed by the DB cluster. For more information, see CreateDBCluster.

Default: false

createDBInstance_dbClusterIdentifier :: Lens' CreateDBInstance (Maybe Text)

The identifier of the DB cluster that the instance will belong to.

For information on creating a DB cluster, see CreateDBCluster.

Type: String

createDBInstance_masterUserPassword :: Lens' CreateDBInstance (Maybe Text)

Not supported by Neptune.

createDBInstance_publiclyAccessible :: Lens' CreateDBInstance (Maybe Bool)

This flag should no longer be used.

createDBInstance_autoMinorVersionUpgrade :: Lens' CreateDBInstance (Maybe Bool)

Indicates that minor engine upgrades are applied automatically to the DB instance during the maintenance window.

Default: true

createDBInstance_masterUsername :: Lens' CreateDBInstance (Maybe Text)

Not supported by Neptune.

createDBInstance_dbSubnetGroupName :: Lens' CreateDBInstance (Maybe Text)

A DB subnet group to associate with this DB instance.

If there is no DB subnet group, then it is a non-VPC DB instance.

createDBInstance_monitoringRoleArn :: Lens' CreateDBInstance (Maybe Text)

The ARN for the IAM role that permits Neptune to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

createDBInstance_iops :: Lens' CreateDBInstance (Maybe Int)

The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance.

createDBInstance_domain :: Lens' CreateDBInstance (Maybe Text)

Specify the Active Directory Domain to create the instance in.

createDBInstance_monitoringInterval :: Lens' CreateDBInstance (Maybe Int)

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance. To disable collecting Enhanced Monitoring metrics, specify 0. The default is 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

createDBInstance_tdeCredentialPassword :: Lens' CreateDBInstance (Maybe Text)

The password for the given ARN from the key store in order to access the device.

createDBInstance_promotionTier :: Lens' CreateDBInstance (Maybe Int)

A value that specifies the order in which an Read Replica is promoted to the primary instance after a failure of the existing primary instance.

Default: 1

Valid Values: 0 - 15

createDBInstance_licenseModel :: Lens' CreateDBInstance (Maybe Text)

License model information for this DB instance.

Valid values: license-included | bring-your-own-license | general-public-license

createDBInstance_preferredMaintenanceWindow :: Lens' CreateDBInstance (Maybe Text)

The time range each week during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Region, occurring on a random day of the week.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

createDBInstance_characterSetName :: Lens' CreateDBInstance (Maybe Text)

(Not supported by Neptune)

createDBInstance_enablePerformanceInsights :: Lens' CreateDBInstance (Maybe Bool)

(Not supported by Neptune)

createDBInstance_kmsKeyId :: Lens' CreateDBInstance (Maybe Text)

The Amazon KMS key identifier for an encrypted DB instance.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are creating a DB instance with the same Amazon account that owns the KMS encryption key used to encrypt the new DB instance, then you can use the KMS key alias instead of the ARN for the KM encryption key.

Not applicable. The KMS key identifier is managed by the DB cluster. For more information, see CreateDBCluster.

If the StorageEncrypted parameter is true, and you do not specify a value for the KmsKeyId parameter, then Amazon Neptune will use your default encryption key. Amazon KMS creates the default encryption key for your Amazon account. Your Amazon account has a different default encryption key for each Amazon Region.

createDBInstance_dbParameterGroupName :: Lens' CreateDBInstance (Maybe Text)

The name of the DB parameter group to associate with this DB instance. If this argument is omitted, the default DBParameterGroup for the specified engine is used.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

createDBInstance_preferredBackupWindow :: Lens' CreateDBInstance (Maybe Text)

The daily time range during which automated backups are created.

Not applicable. The daily time range for creating automated backups is managed by the DB cluster. For more information, see CreateDBCluster.

createDBInstance_availabilityZone :: Lens' CreateDBInstance (Maybe Text)

The EC2 Availability Zone that the DB instance is created in

Default: A random, system-chosen Availability Zone in the endpoint's Amazon Region.

Example: us-east-1d

Constraint: The AvailabilityZone parameter can't be specified if the MultiAZ parameter is set to true. The specified Availability Zone must be in the same Amazon Region as the current endpoint.

createDBInstance_backupRetentionPeriod :: Lens' CreateDBInstance (Maybe Int)

The number of days for which automated backups are retained.

Not applicable. The retention period for automated backups is managed by the DB cluster. For more information, see CreateDBCluster.

Default: 1

Constraints:

Must be a value from 0 to 35

Cannot be set to 0 if the DB instance is a source to Read Replicas

createDBInstance_performanceInsightsKMSKeyId :: Lens' CreateDBInstance (Maybe Text)

(Not supported by Neptune)

createDBInstance_vpcSecurityGroupIds :: Lens' CreateDBInstance (Maybe [Text])

A list of EC2 VPC security groups to associate with this DB instance.

Not applicable. The associated list of EC2 VPC security groups is managed by the DB cluster. For more information, see CreateDBCluster.

Default: The default EC2 VPC security group for the DB subnet group's VPC.

createDBInstance_multiAZ :: Lens' CreateDBInstance (Maybe Bool)

Specifies if the DB instance is a Multi-AZ deployment. You can't set the AvailabilityZone parameter if the MultiAZ parameter is set to true.

createDBInstance_allocatedStorage :: Lens' CreateDBInstance (Maybe Int)

Not supported by Neptune.

createDBInstance_optionGroupName :: Lens' CreateDBInstance (Maybe Text)

(Not supported by Neptune)

createDBInstance_copyTagsToSnapshot :: Lens' CreateDBInstance (Maybe Bool)

True to copy all tags from the DB instance to snapshots of the DB instance, and otherwise false. The default is false.

createDBInstance_timezone :: Lens' CreateDBInstance (Maybe Text)

The time zone of the DB instance.

createDBInstance_tdeCredentialArn :: Lens' CreateDBInstance (Maybe Text)

The ARN from the key store with which to associate the instance for TDE encryption.

createDBInstance_domainIAMRoleName :: Lens' CreateDBInstance (Maybe Text)

Specify the name of the IAM role to be used when making API calls to the Directory Service.

createDBInstance_tags :: Lens' CreateDBInstance (Maybe [Tag])

The tags to assign to the new instance.

createDBInstance_port :: Lens' CreateDBInstance (Maybe Int)

The port number on which the database accepts connections.

Not applicable. The port is managed by the DB cluster. For more information, see CreateDBCluster.

Default: 8182

Type: Integer

createDBInstance_enableIAMDatabaseAuthentication :: Lens' CreateDBInstance (Maybe Bool)

Not supported by Neptune (ignored).

createDBInstance_storageType :: Lens' CreateDBInstance (Maybe Text)

Specifies the storage type to be associated with the DB instance.

Not applicable. Storage is managed by the DB Cluster.

createDBInstance_enableCloudwatchLogsExports :: Lens' CreateDBInstance (Maybe [Text])

The list of log types that need to be enabled for exporting to CloudWatch Logs.

createDBInstance_dbName :: Lens' CreateDBInstance (Maybe Text)

Not supported.

createDBInstance_dbInstanceIdentifier :: Lens' CreateDBInstance Text

The DB instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

createDBInstance_dbInstanceClass :: Lens' CreateDBInstance Text

The compute and memory capacity of the DB instance, for example, db.m4.large. Not all DB instance classes are available in all Amazon Regions.

createDBInstance_engine :: Lens' CreateDBInstance Text

The name of the database engine to be used for this instance.

Valid Values: neptune

Destructuring the Response

data CreateDBInstanceResponse

See: newCreateDBInstanceResponse smart constructor.

Constructors

CreateDBInstanceResponse'

Fields

dbInstance :: Maybe DBInstance

httpStatus :: Int

The response's http status code.

newCreateDBInstanceResponse

Arguments

:: Int
$sel:httpStatus:CreateDBInstanceResponse'

-> CreateDBInstanceResponse

Create a value of CreateDBInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbInstance:CreateDBInstanceResponse', createDBInstanceResponse_dbInstance - Undocumented member.

$sel:httpStatus:CreateDBInstanceResponse', createDBInstanceResponse_httpStatus - The response's http status code.

Response Lenses

createDBInstanceResponse_dbInstance :: Lens' CreateDBInstanceResponse (Maybe DBInstance)

Undocumented member.

createDBInstanceResponse_httpStatus :: Lens' CreateDBInstanceResponse Int

The response's http status code.