Network.AWS.Neptune.DescribeEventCategories

Description

Displays a list of categories for all event source types, or, if specified, for a specified source type.

Creating a Request

data DescribeEventCategories

See: newDescribeEventCategories smart constructor.

Constructors

DescribeEventCategories'

Fields

sourceType :: Maybe Text

The type of source that is generating the events.

Valid values: db-instance | db-parameter-group | db-security-group | db-snapshot

filters :: Maybe [Filter]

This parameter is not currently supported.

newDescribeEventCategories :: DescribeEventCategories

Create a value of DescribeEventCategories with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceType:DescribeEventCategories', describeEventCategories_sourceType - The type of source that is generating the events.

Valid values: db-instance | db-parameter-group | db-security-group | db-snapshot

$sel:filters:DescribeEventCategories', describeEventCategories_filters - This parameter is not currently supported.

Request Lenses

describeEventCategories_sourceType :: Lens' DescribeEventCategories (Maybe Text)

The type of source that is generating the events.

Valid values: db-instance | db-parameter-group | db-security-group | db-snapshot

describeEventCategories_filters :: Lens' DescribeEventCategories (Maybe [Filter])

This parameter is not currently supported.

Destructuring the Response

data DescribeEventCategoriesResponse

See: newDescribeEventCategoriesResponse smart constructor.

Constructors

DescribeEventCategoriesResponse'

Fields

eventCategoriesMapList :: Maybe [EventCategoriesMap]

A list of EventCategoriesMap data types.

httpStatus :: Int

The response's http status code.

newDescribeEventCategoriesResponse

Arguments

:: Int
$sel:httpStatus:DescribeEventCategoriesResponse'

-> DescribeEventCategoriesResponse

Create a value of DescribeEventCategoriesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventCategoriesMapList:DescribeEventCategoriesResponse', describeEventCategoriesResponse_eventCategoriesMapList - A list of EventCategoriesMap data types.

$sel:httpStatus:DescribeEventCategoriesResponse', describeEventCategoriesResponse_httpStatus - The response's http status code.

Response Lenses

describeEventCategoriesResponse_eventCategoriesMapList :: Lens' DescribeEventCategoriesResponse (Maybe [EventCategoriesMap])

A list of EventCategoriesMap data types.

describeEventCategoriesResponse_httpStatus :: Lens' DescribeEventCategoriesResponse Int

The response's http status code.