Network.AWS.Neptune.DescribeEvents

Description

Returns events related to DB instances, DB security groups, DB snapshots, and DB parameter groups for the past 14 days. Events specific to a particular DB instance, DB security group, database snapshot, or DB parameter group can be obtained by providing the name as a parameter. By default, the past hour of events are returned.

This operation returns paginated results.

Creating a Request

data DescribeEvents

See: newDescribeEvents smart constructor.

Constructors

DescribeEvents'

Fields

startTime :: Maybe ISO8601

The beginning of the time interval to retrieve events for, specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page.

Example: 2009-07-08T18:00Z

sourceType :: Maybe SourceType

The event source to retrieve events for. If no value is specified, all events are returned.

filters :: Maybe [Filter]

This parameter is not currently supported.

sourceIdentifier :: Maybe Text

The identifier of the event source for which events are returned. If not specified, then all sources are included in the response.

Constraints:

If SourceIdentifier is supplied, SourceType must also be provided.

If the source type is DBInstance, then a DBInstanceIdentifier must be supplied.

If the source type is DBSecurityGroup, a DBSecurityGroupName must be supplied.

If the source type is DBParameterGroup, a DBParameterGroupName must be supplied.

If the source type is DBSnapshot, a DBSnapshotIdentifier must be supplied.

Cannot end with a hyphen or contain two consecutive hyphens.

eventCategories :: Maybe [Text]

A list of event categories that trigger notifications for a event notification subscription.

marker :: Maybe Text

An optional pagination token provided by a previous DescribeEvents request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

endTime :: Maybe ISO8601

The end of the time interval for which to retrieve events, specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page.

Example: 2009-07-08T18:00Z

duration :: Maybe Int

The number of minutes to retrieve events for.

Default: 60

newDescribeEvents :: DescribeEvents

Create a value of DescribeEvents with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:startTime:DescribeEvents', describeEvents_startTime - The beginning of the time interval to retrieve events for, specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page.

Example: 2009-07-08T18:00Z

$sel:sourceType:DescribeEvents', describeEvents_sourceType - The event source to retrieve events for. If no value is specified, all events are returned.

$sel:filters:DescribeEvents', describeEvents_filters - This parameter is not currently supported.

$sel:sourceIdentifier:DescribeEvents', describeEvents_sourceIdentifier - The identifier of the event source for which events are returned. If not specified, then all sources are included in the response.

Constraints:

If SourceIdentifier is supplied, SourceType must also be provided.

If the source type is DBInstance, then a DBInstanceIdentifier must be supplied.

If the source type is DBSecurityGroup, a DBSecurityGroupName must be supplied.

If the source type is DBParameterGroup, a DBParameterGroupName must be supplied.

If the source type is DBSnapshot, a DBSnapshotIdentifier must be supplied.

Cannot end with a hyphen or contain two consecutive hyphens.

$sel:eventCategories:DescribeEvents', describeEvents_eventCategories - A list of event categories that trigger notifications for a event notification subscription.

$sel:marker:DescribeEvents', describeEvents_marker - An optional pagination token provided by a previous DescribeEvents request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:maxRecords:DescribeEvents', describeEvents_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

$sel:endTime:DescribeEvents', describeEvents_endTime - The end of the time interval for which to retrieve events, specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page.

Example: 2009-07-08T18:00Z

$sel:duration:DescribeEvents', describeEvents_duration - The number of minutes to retrieve events for.

Default: 60

Request Lenses

describeEvents_startTime :: Lens' DescribeEvents (Maybe UTCTime)

The beginning of the time interval to retrieve events for, specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page.

Example: 2009-07-08T18:00Z

describeEvents_sourceType :: Lens' DescribeEvents (Maybe SourceType)

The event source to retrieve events for. If no value is specified, all events are returned.

describeEvents_filters :: Lens' DescribeEvents (Maybe [Filter])

This parameter is not currently supported.

describeEvents_sourceIdentifier :: Lens' DescribeEvents (Maybe Text)

The identifier of the event source for which events are returned. If not specified, then all sources are included in the response.

Constraints:

If SourceIdentifier is supplied, SourceType must also be provided.

If the source type is DBInstance, then a DBInstanceIdentifier must be supplied.

If the source type is DBSecurityGroup, a DBSecurityGroupName must be supplied.

If the source type is DBParameterGroup, a DBParameterGroupName must be supplied.

If the source type is DBSnapshot, a DBSnapshotIdentifier must be supplied.

Cannot end with a hyphen or contain two consecutive hyphens.

describeEvents_eventCategories :: Lens' DescribeEvents (Maybe [Text])

A list of event categories that trigger notifications for a event notification subscription.

describeEvents_marker :: Lens' DescribeEvents (Maybe Text)

An optional pagination token provided by a previous DescribeEvents request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeEvents_maxRecords :: Lens' DescribeEvents (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

Default: 100

Constraints: Minimum 20, maximum 100.

describeEvents_endTime :: Lens' DescribeEvents (Maybe UTCTime)

The end of the time interval for which to retrieve events, specified in ISO 8601 format. For more information about ISO 8601, go to the ISO8601 Wikipedia page.

Example: 2009-07-08T18:00Z

describeEvents_duration :: Lens' DescribeEvents (Maybe Int)

The number of minutes to retrieve events for.

Default: 60

Destructuring the Response

data DescribeEventsResponse

See: newDescribeEventsResponse smart constructor.

Constructors

DescribeEventsResponse'

Fields

events :: Maybe [Event]

A list of Event instances.

marker :: Maybe Text

An optional pagination token provided by a previous Events request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

httpStatus :: Int

The response's http status code.

newDescribeEventsResponse

Arguments

:: Int
$sel:httpStatus:DescribeEventsResponse'

-> DescribeEventsResponse

Create a value of DescribeEventsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:events:DescribeEventsResponse', describeEventsResponse_events - A list of Event instances.

$sel:marker:DescribeEvents', describeEventsResponse_marker - An optional pagination token provided by a previous Events request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

$sel:httpStatus:DescribeEventsResponse', describeEventsResponse_httpStatus - The response's http status code.

Response Lenses

describeEventsResponse_events :: Lens' DescribeEventsResponse (Maybe [Event])

A list of Event instances.

describeEventsResponse_marker :: Lens' DescribeEventsResponse (Maybe Text)

An optional pagination token provided by a previous Events request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

describeEventsResponse_httpStatus :: Lens' DescribeEventsResponse Int

The response's http status code.