Network.AWS.Neptune.FailoverDBCluster

Description

Forces a failover for a DB cluster.

A failover for a DB cluster promotes one of the Read Replicas (read-only instances) in the DB cluster to be the primary instance (the cluster writer).

Amazon Neptune will automatically fail over to a Read Replica, if one exists, when the primary instance fails. You can force a failover when you want to simulate a failure of a primary instance for testing. Because each instance in a DB cluster has its own endpoint address, you will need to clean up and re-establish any existing connections that use those endpoint addresses when the failover is complete.

Creating a Request

data FailoverDBCluster

See: newFailoverDBCluster smart constructor.

Constructors

FailoverDBCluster'

Fields

dbClusterIdentifier :: Maybe Text

A DB cluster identifier to force a failover for. This parameter is not case-sensitive.

Constraints:

Must match the identifier of an existing DBCluster.

targetDBInstanceIdentifier :: Maybe Text

The name of the instance to promote to the primary instance.

You must specify the instance identifier for an Read Replica in the DB cluster. For example, mydbcluster-replica1.

newFailoverDBCluster :: FailoverDBCluster

Create a value of FailoverDBCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbClusterIdentifier:FailoverDBCluster', failoverDBCluster_dbClusterIdentifier - A DB cluster identifier to force a failover for. This parameter is not case-sensitive.

Constraints:

Must match the identifier of an existing DBCluster.

$sel:targetDBInstanceIdentifier:FailoverDBCluster', failoverDBCluster_targetDBInstanceIdentifier - The name of the instance to promote to the primary instance.

You must specify the instance identifier for an Read Replica in the DB cluster. For example, mydbcluster-replica1.

Request Lenses

failoverDBCluster_dbClusterIdentifier :: Lens' FailoverDBCluster (Maybe Text)

A DB cluster identifier to force a failover for. This parameter is not case-sensitive.

Constraints:

Must match the identifier of an existing DBCluster.

failoverDBCluster_targetDBInstanceIdentifier :: Lens' FailoverDBCluster (Maybe Text)

The name of the instance to promote to the primary instance.

You must specify the instance identifier for an Read Replica in the DB cluster. For example, mydbcluster-replica1.

Destructuring the Response

data FailoverDBClusterResponse

See: newFailoverDBClusterResponse smart constructor.

Constructors

FailoverDBClusterResponse'

Fields

dbCluster :: Maybe DBCluster

httpStatus :: Int

The response's http status code.

newFailoverDBClusterResponse

Arguments

:: Int
$sel:httpStatus:FailoverDBClusterResponse'

-> FailoverDBClusterResponse

Create a value of FailoverDBClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbCluster:FailoverDBClusterResponse', failoverDBClusterResponse_dbCluster - Undocumented member.

$sel:httpStatus:FailoverDBClusterResponse', failoverDBClusterResponse_httpStatus - The response's http status code.

Response Lenses

failoverDBClusterResponse_dbCluster :: Lens' FailoverDBClusterResponse (Maybe DBCluster)

Undocumented member.

failoverDBClusterResponse_httpStatus :: Lens' FailoverDBClusterResponse Int

The response's http status code.