Network.AWS.Neptune.ModifyDBInstance

Description

Modifies settings for a DB instance. You can change one or more database configuration parameters by specifying these parameters and the new values in the request. To learn what modifications you can make to your DB instance, call DescribeValidDBInstanceModifications before you call ModifyDBInstance.

Creating a Request

data ModifyDBInstance

See: newModifyDBInstance smart constructor.

Constructors

ModifyDBInstance'

Fields

engineVersion :: Maybe Text

The version number of the database engine to upgrade to. Currently, setting this parameter has no effect. To upgrade your database engine to the most recent release, use the ApplyPendingMaintenanceAction API.

dbSecurityGroups :: Maybe [Text]

A list of DB security groups to authorize on this DB instance. Changing this setting doesn't result in an outage and the change is asynchronously applied as soon as possible.

Constraints:

If supplied, must match existing DBSecurityGroups.

dbPortNumber :: Maybe Int

The port number on which the database accepts connections.

The value of the DBPortNumber parameter must not match any of the port values specified for options in the option group for the DB instance.

Your database will restart when you change the DBPortNumber value regardless of the value of the ApplyImmediately parameter.

Default: 8182

deletionProtection :: Maybe Bool

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. See Deleting a DB Instance.

masterUserPassword :: Maybe Text

Not supported by Neptune.

publiclyAccessible :: Maybe Bool

This flag should no longer be used.

autoMinorVersionUpgrade :: Maybe Bool

Indicates that minor version upgrades are applied automatically to the DB instance during the maintenance window. Changing this parameter doesn't result in an outage except in the following case and the change is asynchronously applied as soon as possible. An outage will result if this parameter is set to true during the maintenance window, and a newer minor version is available, and Neptune has enabled auto patching for that engine version.

dbSubnetGroupName :: Maybe Text

The new DB subnet group for the DB instance. You can use this parameter to move your DB instance to a different VPC.

Changing the subnet group causes an outage during the change. The change is applied during the next maintenance window, unless you specify true for the ApplyImmediately parameter.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetGroup

monitoringRoleArn :: Maybe Text

The ARN for the IAM role that permits Neptune to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

iops :: Maybe Int

The new Provisioned IOPS (I/O operations per second) value for the instance.

Changing this setting doesn't result in an outage and the change is applied during the next maintenance window unless the ApplyImmediately parameter is set to true for this request.

Default: Uses existing setting

allowMajorVersionUpgrade :: Maybe Bool

Indicates that major version upgrades are allowed. Changing this parameter doesn't result in an outage and the change is asynchronously applied as soon as possible.

newDBInstanceIdentifier' :: Maybe Text

The new DB instance identifier for the DB instance when renaming a DB instance. When you change the DB instance identifier, an instance reboot will occur immediately if you set Apply Immediately to true, or will occur during the next maintenance window if Apply Immediately to false. This value is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

The first character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

domain :: Maybe Text

Not supported.

monitoringInterval :: Maybe Int

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance. To disable collecting Enhanced Monitoring metrics, specify 0. The default is 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

cloudwatchLogsExportConfiguration :: Maybe CloudwatchLogsExportConfiguration

The configuration setting for the log types to be enabled for export to CloudWatch Logs for a specific DB instance or DB cluster.

tdeCredentialPassword :: Maybe Text

The password for the given ARN from the key store in order to access the device.

dbInstanceClass :: Maybe Text

The new compute and memory capacity of the DB instance, for example, db.m4.large. Not all DB instance classes are available in all AWS Regions.

If you modify the DB instance class, an outage occurs during the change. The change is applied during the next maintenance window, unless ApplyImmediately is specified as true for this request.

Default: Uses existing setting

promotionTier :: Maybe Int

A value that specifies the order in which a Read Replica is promoted to the primary instance after a failure of the existing primary instance.

Default: 1

Valid Values: 0 - 15

licenseModel :: Maybe Text

Not supported by Neptune.

preferredMaintenanceWindow :: Maybe Text

The weekly time range (in UTC) during which system maintenance can occur, which might result in an outage. Changing this parameter doesn't result in an outage, except in the following situation, and the change is asynchronously applied as soon as possible. If there are pending actions that cause a reboot, and the maintenance window is changed to include the current time, then changing this parameter will cause a reboot of the DB instance. If moving this window to the current time, there must be at least 30 minutes between the current time and end of the window to ensure pending changes are applied.

Default: Uses existing setting

Format: ddd:hh24:mi-ddd:hh24:mi

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Must be at least 30 minutes

cACertificateIdentifier :: Maybe Text

Indicates the certificate that needs to be associated with the instance.

enablePerformanceInsights :: Maybe Bool

(Not supported by Neptune)

dbParameterGroupName :: Maybe Text

The name of the DB parameter group to apply to the DB instance. Changing this setting doesn't result in an outage. The parameter group name itself is changed immediately, but the actual parameter changes are not applied until you reboot the instance without failover. The db instance will NOT be rebooted automatically and the parameter changes will NOT be applied during the next maintenance window.

Default: Uses existing setting

Constraints: The DB parameter group must be in the same DB parameter group family as this DB instance.

preferredBackupWindow :: Maybe Text

The daily time range during which automated backups are created if automated backups are enabled.

Not applicable. The daily time range for creating automated backups is managed by the DB cluster. For more information, see ModifyDBCluster.

Constraints:

Must be in the format hh24:mi-hh24:mi

Must be in Universal Time Coordinated (UTC)

Must not conflict with the preferred maintenance window

Must be at least 30 minutes

backupRetentionPeriod :: Maybe Int

Not applicable. The retention period for automated backups is managed by the DB cluster. For more information, see ModifyDBCluster.

Default: Uses existing setting

performanceInsightsKMSKeyId :: Maybe Text

(Not supported by Neptune)

vpcSecurityGroupIds :: Maybe [Text]

A list of EC2 VPC security groups to authorize on this DB instance. This change is asynchronously applied as soon as possible.

Not applicable. The associated list of EC2 VPC security groups is managed by the DB cluster. For more information, see ModifyDBCluster.

Constraints:

If supplied, must match existing VpcSecurityGroupIds.

multiAZ :: Maybe Bool

Specifies if the DB instance is a Multi-AZ deployment. Changing this parameter doesn't result in an outage and the change is applied during the next maintenance window unless the ApplyImmediately parameter is set to true for this request.

allocatedStorage :: Maybe Int

Not supported by Neptune.

applyImmediately :: Maybe Bool

Specifies whether the modifications in this request and any pending modifications are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the DB instance.

If this parameter is set to false, changes to the DB instance are applied during the next maintenance window. Some parameter changes can cause an outage and are applied on the next call to RebootDBInstance, or the next failure reboot.

Default: false

optionGroupName :: Maybe Text

(Not supported by Neptune)

copyTagsToSnapshot :: Maybe Bool

True to copy all tags from the DB instance to snapshots of the DB instance, and otherwise false. The default is false.

tdeCredentialArn :: Maybe Text

The ARN from the key store with which to associate the instance for TDE encryption.

domainIAMRoleName :: Maybe Text

Not supported

enableIAMDatabaseAuthentication :: Maybe Bool

True to enable mapping of Amazon Identity and Access Management (IAM) accounts to database accounts, and otherwise false.

You can enable IAM database authentication for the following database engines

Not applicable. Mapping Amazon IAM accounts to database accounts is managed by the DB cluster. For more information, see ModifyDBCluster.

Default: false

storageType :: Maybe Text

Not supported.

dbInstanceIdentifier :: Text

The DB instance identifier. This value is stored as a lowercase string.

Constraints:

Must match the identifier of an existing DBInstance.

newModifyDBInstance

Arguments

:: Text
$sel:dbInstanceIdentifier:ModifyDBInstance'

-> ModifyDBInstance

Create a value of ModifyDBInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:ModifyDBInstance', modifyDBInstance_engineVersion - The version number of the database engine to upgrade to. Currently, setting this parameter has no effect. To upgrade your database engine to the most recent release, use the ApplyPendingMaintenanceAction API.

$sel:dbSecurityGroups:ModifyDBInstance', modifyDBInstance_dbSecurityGroups - A list of DB security groups to authorize on this DB instance. Changing this setting doesn't result in an outage and the change is asynchronously applied as soon as possible.

Constraints:

If supplied, must match existing DBSecurityGroups.

$sel:dbPortNumber:ModifyDBInstance', modifyDBInstance_dbPortNumber - The port number on which the database accepts connections.

The value of the DBPortNumber parameter must not match any of the port values specified for options in the option group for the DB instance.

Your database will restart when you change the DBPortNumber value regardless of the value of the ApplyImmediately parameter.

Default: 8182

$sel:deletionProtection:ModifyDBInstance', modifyDBInstance_deletionProtection - A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. See Deleting a DB Instance.

$sel:masterUserPassword:ModifyDBInstance', modifyDBInstance_masterUserPassword - Not supported by Neptune.

$sel:publiclyAccessible:ModifyDBInstance', modifyDBInstance_publiclyAccessible - This flag should no longer be used.

$sel:autoMinorVersionUpgrade:ModifyDBInstance', modifyDBInstance_autoMinorVersionUpgrade - Indicates that minor version upgrades are applied automatically to the DB instance during the maintenance window. Changing this parameter doesn't result in an outage except in the following case and the change is asynchronously applied as soon as possible. An outage will result if this parameter is set to true during the maintenance window, and a newer minor version is available, and Neptune has enabled auto patching for that engine version.

$sel:dbSubnetGroupName:ModifyDBInstance', modifyDBInstance_dbSubnetGroupName - The new DB subnet group for the DB instance. You can use this parameter to move your DB instance to a different VPC.

Changing the subnet group causes an outage during the change. The change is applied during the next maintenance window, unless you specify true for the ApplyImmediately parameter.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetGroup

$sel:monitoringRoleArn:ModifyDBInstance', modifyDBInstance_monitoringRoleArn - The ARN for the IAM role that permits Neptune to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

$sel:iops:ModifyDBInstance', modifyDBInstance_iops - The new Provisioned IOPS (I/O operations per second) value for the instance.

Changing this setting doesn't result in an outage and the change is applied during the next maintenance window unless the ApplyImmediately parameter is set to true for this request.

Default: Uses existing setting

$sel:allowMajorVersionUpgrade:ModifyDBInstance', modifyDBInstance_allowMajorVersionUpgrade - Indicates that major version upgrades are allowed. Changing this parameter doesn't result in an outage and the change is asynchronously applied as soon as possible.

$sel:newDBInstanceIdentifier':ModifyDBInstance', modifyDBInstance_newDBInstanceIdentifier - The new DB instance identifier for the DB instance when renaming a DB instance. When you change the DB instance identifier, an instance reboot will occur immediately if you set Apply Immediately to true, or will occur during the next maintenance window if Apply Immediately to false. This value is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

The first character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

$sel:domain:ModifyDBInstance', modifyDBInstance_domain - Not supported.

$sel:monitoringInterval:ModifyDBInstance', modifyDBInstance_monitoringInterval - The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance. To disable collecting Enhanced Monitoring metrics, specify 0. The default is 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

$sel:cloudwatchLogsExportConfiguration:ModifyDBInstance', modifyDBInstance_cloudwatchLogsExportConfiguration - The configuration setting for the log types to be enabled for export to CloudWatch Logs for a specific DB instance or DB cluster.

$sel:tdeCredentialPassword:ModifyDBInstance', modifyDBInstance_tdeCredentialPassword - The password for the given ARN from the key store in order to access the device.

$sel:dbInstanceClass:ModifyDBInstance', modifyDBInstance_dbInstanceClass - The new compute and memory capacity of the DB instance, for example, db.m4.large. Not all DB instance classes are available in all AWS Regions.

If you modify the DB instance class, an outage occurs during the change. The change is applied during the next maintenance window, unless ApplyImmediately is specified as true for this request.

Default: Uses existing setting

$sel:promotionTier:ModifyDBInstance', modifyDBInstance_promotionTier - A value that specifies the order in which a Read Replica is promoted to the primary instance after a failure of the existing primary instance.

Default: 1

Valid Values: 0 - 15

$sel:licenseModel:ModifyDBInstance', modifyDBInstance_licenseModel - Not supported by Neptune.

$sel:preferredMaintenanceWindow:ModifyDBInstance', modifyDBInstance_preferredMaintenanceWindow - The weekly time range (in UTC) during which system maintenance can occur, which might result in an outage. Changing this parameter doesn't result in an outage, except in the following situation, and the change is asynchronously applied as soon as possible. If there are pending actions that cause a reboot, and the maintenance window is changed to include the current time, then changing this parameter will cause a reboot of the DB instance. If moving this window to the current time, there must be at least 30 minutes between the current time and end of the window to ensure pending changes are applied.

Default: Uses existing setting

Format: ddd:hh24:mi-ddd:hh24:mi

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Must be at least 30 minutes

$sel:cACertificateIdentifier:ModifyDBInstance', modifyDBInstance_cACertificateIdentifier - Indicates the certificate that needs to be associated with the instance.

$sel:enablePerformanceInsights:ModifyDBInstance', modifyDBInstance_enablePerformanceInsights - (Not supported by Neptune)

$sel:dbParameterGroupName:ModifyDBInstance', modifyDBInstance_dbParameterGroupName - The name of the DB parameter group to apply to the DB instance. Changing this setting doesn't result in an outage. The parameter group name itself is changed immediately, but the actual parameter changes are not applied until you reboot the instance without failover. The db instance will NOT be rebooted automatically and the parameter changes will NOT be applied during the next maintenance window.

Default: Uses existing setting

Constraints: The DB parameter group must be in the same DB parameter group family as this DB instance.

$sel:preferredBackupWindow:ModifyDBInstance', modifyDBInstance_preferredBackupWindow - The daily time range during which automated backups are created if automated backups are enabled.

Not applicable. The daily time range for creating automated backups is managed by the DB cluster. For more information, see ModifyDBCluster.

Constraints:

Must be in the format hh24:mi-hh24:mi

Must be in Universal Time Coordinated (UTC)

Must not conflict with the preferred maintenance window

Must be at least 30 minutes

$sel:backupRetentionPeriod:ModifyDBInstance', modifyDBInstance_backupRetentionPeriod - Not applicable. The retention period for automated backups is managed by the DB cluster. For more information, see ModifyDBCluster.

Default: Uses existing setting

$sel:performanceInsightsKMSKeyId:ModifyDBInstance', modifyDBInstance_performanceInsightsKMSKeyId - (Not supported by Neptune)

$sel:vpcSecurityGroupIds:ModifyDBInstance', modifyDBInstance_vpcSecurityGroupIds - A list of EC2 VPC security groups to authorize on this DB instance. This change is asynchronously applied as soon as possible.

Not applicable. The associated list of EC2 VPC security groups is managed by the DB cluster. For more information, see ModifyDBCluster.

Constraints:

If supplied, must match existing VpcSecurityGroupIds.

$sel:multiAZ:ModifyDBInstance', modifyDBInstance_multiAZ - Specifies if the DB instance is a Multi-AZ deployment. Changing this parameter doesn't result in an outage and the change is applied during the next maintenance window unless the ApplyImmediately parameter is set to true for this request.

$sel:allocatedStorage:ModifyDBInstance', modifyDBInstance_allocatedStorage - Not supported by Neptune.

$sel:applyImmediately:ModifyDBInstance', modifyDBInstance_applyImmediately - Specifies whether the modifications in this request and any pending modifications are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the DB instance.

If this parameter is set to false, changes to the DB instance are applied during the next maintenance window. Some parameter changes can cause an outage and are applied on the next call to RebootDBInstance, or the next failure reboot.

Default: false

$sel:optionGroupName:ModifyDBInstance', modifyDBInstance_optionGroupName - (Not supported by Neptune)

$sel:copyTagsToSnapshot:ModifyDBInstance', modifyDBInstance_copyTagsToSnapshot - True to copy all tags from the DB instance to snapshots of the DB instance, and otherwise false. The default is false.

$sel:tdeCredentialArn:ModifyDBInstance', modifyDBInstance_tdeCredentialArn - The ARN from the key store with which to associate the instance for TDE encryption.

$sel:domainIAMRoleName:ModifyDBInstance', modifyDBInstance_domainIAMRoleName - Not supported

$sel:enableIAMDatabaseAuthentication:ModifyDBInstance', modifyDBInstance_enableIAMDatabaseAuthentication - True to enable mapping of Amazon Identity and Access Management (IAM) accounts to database accounts, and otherwise false.

You can enable IAM database authentication for the following database engines

Not applicable. Mapping Amazon IAM accounts to database accounts is managed by the DB cluster. For more information, see ModifyDBCluster.

Default: false

$sel:storageType:ModifyDBInstance', modifyDBInstance_storageType - Not supported.

$sel:dbInstanceIdentifier:ModifyDBInstance', modifyDBInstance_dbInstanceIdentifier - The DB instance identifier. This value is stored as a lowercase string.

Constraints:

Must match the identifier of an existing DBInstance.

Request Lenses

modifyDBInstance_engineVersion :: Lens' ModifyDBInstance (Maybe Text)

The version number of the database engine to upgrade to. Currently, setting this parameter has no effect. To upgrade your database engine to the most recent release, use the ApplyPendingMaintenanceAction API.

modifyDBInstance_dbSecurityGroups :: Lens' ModifyDBInstance (Maybe [Text])

A list of DB security groups to authorize on this DB instance. Changing this setting doesn't result in an outage and the change is asynchronously applied as soon as possible.

Constraints:

If supplied, must match existing DBSecurityGroups.

modifyDBInstance_dbPortNumber :: Lens' ModifyDBInstance (Maybe Int)

The port number on which the database accepts connections.

The value of the DBPortNumber parameter must not match any of the port values specified for options in the option group for the DB instance.

Your database will restart when you change the DBPortNumber value regardless of the value of the ApplyImmediately parameter.

Default: 8182

modifyDBInstance_deletionProtection :: Lens' ModifyDBInstance (Maybe Bool)

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. See Deleting a DB Instance.

modifyDBInstance_masterUserPassword :: Lens' ModifyDBInstance (Maybe Text)

Not supported by Neptune.

modifyDBInstance_publiclyAccessible :: Lens' ModifyDBInstance (Maybe Bool)

This flag should no longer be used.

modifyDBInstance_autoMinorVersionUpgrade :: Lens' ModifyDBInstance (Maybe Bool)

Indicates that minor version upgrades are applied automatically to the DB instance during the maintenance window. Changing this parameter doesn't result in an outage except in the following case and the change is asynchronously applied as soon as possible. An outage will result if this parameter is set to true during the maintenance window, and a newer minor version is available, and Neptune has enabled auto patching for that engine version.

modifyDBInstance_dbSubnetGroupName :: Lens' ModifyDBInstance (Maybe Text)

The new DB subnet group for the DB instance. You can use this parameter to move your DB instance to a different VPC.

Changing the subnet group causes an outage during the change. The change is applied during the next maintenance window, unless you specify true for the ApplyImmediately parameter.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetGroup

modifyDBInstance_monitoringRoleArn :: Lens' ModifyDBInstance (Maybe Text)

The ARN for the IAM role that permits Neptune to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

modifyDBInstance_iops :: Lens' ModifyDBInstance (Maybe Int)

The new Provisioned IOPS (I/O operations per second) value for the instance.

Changing this setting doesn't result in an outage and the change is applied during the next maintenance window unless the ApplyImmediately parameter is set to true for this request.

Default: Uses existing setting

modifyDBInstance_allowMajorVersionUpgrade :: Lens' ModifyDBInstance (Maybe Bool)

Indicates that major version upgrades are allowed. Changing this parameter doesn't result in an outage and the change is asynchronously applied as soon as possible.

modifyDBInstance_newDBInstanceIdentifier :: Lens' ModifyDBInstance (Maybe Text)

The new DB instance identifier for the DB instance when renaming a DB instance. When you change the DB instance identifier, an instance reboot will occur immediately if you set Apply Immediately to true, or will occur during the next maintenance window if Apply Immediately to false. This value is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

The first character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

modifyDBInstance_domain :: Lens' ModifyDBInstance (Maybe Text)

Not supported.

modifyDBInstance_monitoringInterval :: Lens' ModifyDBInstance (Maybe Int)

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance. To disable collecting Enhanced Monitoring metrics, specify 0. The default is 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

modifyDBInstance_cloudwatchLogsExportConfiguration :: Lens' ModifyDBInstance (Maybe CloudwatchLogsExportConfiguration)

The configuration setting for the log types to be enabled for export to CloudWatch Logs for a specific DB instance or DB cluster.

modifyDBInstance_tdeCredentialPassword :: Lens' ModifyDBInstance (Maybe Text)

The password for the given ARN from the key store in order to access the device.

modifyDBInstance_dbInstanceClass :: Lens' ModifyDBInstance (Maybe Text)

The new compute and memory capacity of the DB instance, for example, db.m4.large. Not all DB instance classes are available in all AWS Regions.

If you modify the DB instance class, an outage occurs during the change. The change is applied during the next maintenance window, unless ApplyImmediately is specified as true for this request.

Default: Uses existing setting

modifyDBInstance_promotionTier :: Lens' ModifyDBInstance (Maybe Int)

A value that specifies the order in which a Read Replica is promoted to the primary instance after a failure of the existing primary instance.

Default: 1

Valid Values: 0 - 15

modifyDBInstance_licenseModel :: Lens' ModifyDBInstance (Maybe Text)

Not supported by Neptune.

modifyDBInstance_preferredMaintenanceWindow :: Lens' ModifyDBInstance (Maybe Text)

The weekly time range (in UTC) during which system maintenance can occur, which might result in an outage. Changing this parameter doesn't result in an outage, except in the following situation, and the change is asynchronously applied as soon as possible. If there are pending actions that cause a reboot, and the maintenance window is changed to include the current time, then changing this parameter will cause a reboot of the DB instance. If moving this window to the current time, there must be at least 30 minutes between the current time and end of the window to ensure pending changes are applied.

Default: Uses existing setting

Format: ddd:hh24:mi-ddd:hh24:mi

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Must be at least 30 minutes

modifyDBInstance_cACertificateIdentifier :: Lens' ModifyDBInstance (Maybe Text)

Indicates the certificate that needs to be associated with the instance.

modifyDBInstance_enablePerformanceInsights :: Lens' ModifyDBInstance (Maybe Bool)

(Not supported by Neptune)

modifyDBInstance_dbParameterGroupName :: Lens' ModifyDBInstance (Maybe Text)

The name of the DB parameter group to apply to the DB instance. Changing this setting doesn't result in an outage. The parameter group name itself is changed immediately, but the actual parameter changes are not applied until you reboot the instance without failover. The db instance will NOT be rebooted automatically and the parameter changes will NOT be applied during the next maintenance window.

Default: Uses existing setting

Constraints: The DB parameter group must be in the same DB parameter group family as this DB instance.

modifyDBInstance_preferredBackupWindow :: Lens' ModifyDBInstance (Maybe Text)

The daily time range during which automated backups are created if automated backups are enabled.

Not applicable. The daily time range for creating automated backups is managed by the DB cluster. For more information, see ModifyDBCluster.

Constraints:

Must be in the format hh24:mi-hh24:mi

Must be in Universal Time Coordinated (UTC)

Must not conflict with the preferred maintenance window

Must be at least 30 minutes

modifyDBInstance_backupRetentionPeriod :: Lens' ModifyDBInstance (Maybe Int)

Not applicable. The retention period for automated backups is managed by the DB cluster. For more information, see ModifyDBCluster.

Default: Uses existing setting

modifyDBInstance_performanceInsightsKMSKeyId :: Lens' ModifyDBInstance (Maybe Text)

(Not supported by Neptune)

modifyDBInstance_vpcSecurityGroupIds :: Lens' ModifyDBInstance (Maybe [Text])

A list of EC2 VPC security groups to authorize on this DB instance. This change is asynchronously applied as soon as possible.

Not applicable. The associated list of EC2 VPC security groups is managed by the DB cluster. For more information, see ModifyDBCluster.

Constraints:

If supplied, must match existing VpcSecurityGroupIds.

modifyDBInstance_multiAZ :: Lens' ModifyDBInstance (Maybe Bool)

Specifies if the DB instance is a Multi-AZ deployment. Changing this parameter doesn't result in an outage and the change is applied during the next maintenance window unless the ApplyImmediately parameter is set to true for this request.

modifyDBInstance_allocatedStorage :: Lens' ModifyDBInstance (Maybe Int)

Not supported by Neptune.

modifyDBInstance_applyImmediately :: Lens' ModifyDBInstance (Maybe Bool)

Specifies whether the modifications in this request and any pending modifications are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the DB instance.

If this parameter is set to false, changes to the DB instance are applied during the next maintenance window. Some parameter changes can cause an outage and are applied on the next call to RebootDBInstance, or the next failure reboot.

Default: false

modifyDBInstance_optionGroupName :: Lens' ModifyDBInstance (Maybe Text)

(Not supported by Neptune)

modifyDBInstance_copyTagsToSnapshot :: Lens' ModifyDBInstance (Maybe Bool)

True to copy all tags from the DB instance to snapshots of the DB instance, and otherwise false. The default is false.

modifyDBInstance_tdeCredentialArn :: Lens' ModifyDBInstance (Maybe Text)

The ARN from the key store with which to associate the instance for TDE encryption.

modifyDBInstance_domainIAMRoleName :: Lens' ModifyDBInstance (Maybe Text)

Not supported

modifyDBInstance_enableIAMDatabaseAuthentication :: Lens' ModifyDBInstance (Maybe Bool)

True to enable mapping of Amazon Identity and Access Management (IAM) accounts to database accounts, and otherwise false.

You can enable IAM database authentication for the following database engines

Not applicable. Mapping Amazon IAM accounts to database accounts is managed by the DB cluster. For more information, see ModifyDBCluster.

Default: false

modifyDBInstance_storageType :: Lens' ModifyDBInstance (Maybe Text)

Not supported.

modifyDBInstance_dbInstanceIdentifier :: Lens' ModifyDBInstance Text

The DB instance identifier. This value is stored as a lowercase string.

Constraints:

Must match the identifier of an existing DBInstance.

Destructuring the Response

data ModifyDBInstanceResponse

See: newModifyDBInstanceResponse smart constructor.

Constructors

ModifyDBInstanceResponse'

Fields

dbInstance :: Maybe DBInstance

httpStatus :: Int

The response's http status code.

newModifyDBInstanceResponse

Arguments

:: Int
$sel:httpStatus:ModifyDBInstanceResponse'

-> ModifyDBInstanceResponse

Create a value of ModifyDBInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbInstance:ModifyDBInstanceResponse', modifyDBInstanceResponse_dbInstance - Undocumented member.

$sel:httpStatus:ModifyDBInstanceResponse', modifyDBInstanceResponse_httpStatus - The response's http status code.

Response Lenses

modifyDBInstanceResponse_dbInstance :: Lens' ModifyDBInstanceResponse (Maybe DBInstance)

Undocumented member.

modifyDBInstanceResponse_httpStatus :: Lens' ModifyDBInstanceResponse Int

The response's http status code.