Network.AWS.Neptune.RestoreDBClusterFromSnapshot

Description

Creates a new DB cluster from a DB snapshot or DB cluster snapshot.

If a DB snapshot is specified, the target DB cluster is created from the source DB snapshot with a default configuration and default security group.

If a DB cluster snapshot is specified, the target DB cluster is created from the source DB cluster restore point with the same configuration as the original source DB cluster, except that the new DB cluster is created with the default security group.

Creating a Request

data RestoreDBClusterFromSnapshot

See: newRestoreDBClusterFromSnapshot smart constructor.

Constructors

RestoreDBClusterFromSnapshot'

Fields

engineVersion :: Maybe Text

The version of the database engine to use for the new DB cluster.

deletionProtection :: Maybe Bool

A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

dbSubnetGroupName :: Maybe Text

The name of the DB subnet group to use for the new DB cluster.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

availabilityZones :: Maybe [Text]

Provides the list of EC2 Availability Zones that instances in the restored DB cluster can be created in.

kmsKeyId :: Maybe Text

The Amazon KMS key identifier to use when restoring an encrypted DB cluster from a DB snapshot or DB cluster snapshot.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are restoring a DB cluster with the same Amazon account that owns the KMS encryption key used to encrypt the new DB cluster, then you can use the KMS key alias instead of the ARN for the KMS encryption key.

If you do not specify a value for the KmsKeyId parameter, then the following will occur:

If the DB snapshot or DB cluster snapshot in SnapshotIdentifier is encrypted, then the restored DB cluster is encrypted using the KMS key that was used to encrypt the DB snapshot or DB cluster snapshot.

If the DB snapshot or DB cluster snapshot in SnapshotIdentifier is not encrypted, then the restored DB cluster is not encrypted.

vpcSecurityGroupIds :: Maybe [Text]

A list of VPC security groups that the new DB cluster will belong to.

databaseName :: Maybe Text

Not supported.

dbClusterParameterGroupName :: Maybe Text

The name of the DB cluster parameter group to associate with the new DB cluster.

Constraints:

If supplied, must match the name of an existing DBClusterParameterGroup.

optionGroupName :: Maybe Text

(Not supported by Neptune)

copyTagsToSnapshot :: Maybe Bool

/If set to true, tags are copied to any snapshot of the restored DB cluster that is created./

tags :: Maybe [Tag]

The tags to be assigned to the restored DB cluster.

port :: Maybe Int

The port number on which the new DB cluster accepts connections.

Constraints: Value must be 1150-65535

Default: The same port as the original DB cluster.

enableIAMDatabaseAuthentication :: Maybe Bool

True to enable mapping of Amazon Identity and Access Management (IAM) accounts to database accounts, and otherwise false.

Default: false

enableCloudwatchLogsExports :: Maybe [Text]

The list of logs that the restored DB cluster is to export to Amazon CloudWatch Logs.

dbClusterIdentifier :: Text

The name of the DB cluster to create from the DB snapshot or DB cluster snapshot. This parameter isn't case-sensitive.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

Example: my-snapshot-id

snapshotIdentifier :: Text

The identifier for the DB snapshot or DB cluster snapshot to restore from.

You can use either the name or the Amazon Resource Name (ARN) to specify a DB cluster snapshot. However, you can use only the ARN to specify a DB snapshot.

Constraints:

Must match the identifier of an existing Snapshot.

engine :: Text

The database engine to use for the new DB cluster.

Default: The same as source

Constraint: Must be compatible with the engine of the source

newRestoreDBClusterFromSnapshot

Arguments

:: Text
$sel:dbClusterIdentifier:RestoreDBClusterFromSnapshot'

-> Text
$sel:snapshotIdentifier:RestoreDBClusterFromSnapshot'

-> Text
$sel:engine:RestoreDBClusterFromSnapshot'

-> RestoreDBClusterFromSnapshot

Create a value of RestoreDBClusterFromSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_engineVersion - The version of the database engine to use for the new DB cluster.

$sel:deletionProtection:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_deletionProtection - A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

$sel:dbSubnetGroupName:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_dbSubnetGroupName - The name of the DB subnet group to use for the new DB cluster.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

$sel:availabilityZones:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_availabilityZones - Provides the list of EC2 Availability Zones that instances in the restored DB cluster can be created in.

$sel:kmsKeyId:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_kmsKeyId - The Amazon KMS key identifier to use when restoring an encrypted DB cluster from a DB snapshot or DB cluster snapshot.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are restoring a DB cluster with the same Amazon account that owns the KMS encryption key used to encrypt the new DB cluster, then you can use the KMS key alias instead of the ARN for the KMS encryption key.

If you do not specify a value for the KmsKeyId parameter, then the following will occur:

If the DB snapshot or DB cluster snapshot in SnapshotIdentifier is encrypted, then the restored DB cluster is encrypted using the KMS key that was used to encrypt the DB snapshot or DB cluster snapshot.

If the DB snapshot or DB cluster snapshot in SnapshotIdentifier is not encrypted, then the restored DB cluster is not encrypted.

$sel:vpcSecurityGroupIds:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_vpcSecurityGroupIds - A list of VPC security groups that the new DB cluster will belong to.

$sel:databaseName:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_databaseName - Not supported.

$sel:dbClusterParameterGroupName:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_dbClusterParameterGroupName - The name of the DB cluster parameter group to associate with the new DB cluster.

Constraints:

If supplied, must match the name of an existing DBClusterParameterGroup.

$sel:optionGroupName:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_optionGroupName - (Not supported by Neptune)

$sel:copyTagsToSnapshot:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_copyTagsToSnapshot - /If set to true, tags are copied to any snapshot of the restored DB cluster that is created./

$sel:tags:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_tags - The tags to be assigned to the restored DB cluster.

$sel:port:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_port - The port number on which the new DB cluster accepts connections.

Constraints: Value must be 1150-65535

Default: The same port as the original DB cluster.

$sel:enableIAMDatabaseAuthentication:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_enableIAMDatabaseAuthentication - True to enable mapping of Amazon Identity and Access Management (IAM) accounts to database accounts, and otherwise false.

Default: false

$sel:enableCloudwatchLogsExports:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_enableCloudwatchLogsExports - The list of logs that the restored DB cluster is to export to Amazon CloudWatch Logs.

$sel:dbClusterIdentifier:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_dbClusterIdentifier - The name of the DB cluster to create from the DB snapshot or DB cluster snapshot. This parameter isn't case-sensitive.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

Example: my-snapshot-id

$sel:snapshotIdentifier:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_snapshotIdentifier - The identifier for the DB snapshot or DB cluster snapshot to restore from.

You can use either the name or the Amazon Resource Name (ARN) to specify a DB cluster snapshot. However, you can use only the ARN to specify a DB snapshot.

Constraints:

Must match the identifier of an existing Snapshot.

$sel:engine:RestoreDBClusterFromSnapshot', restoreDBClusterFromSnapshot_engine - The database engine to use for the new DB cluster.

Default: The same as source

Constraint: Must be compatible with the engine of the source

Request Lenses

restoreDBClusterFromSnapshot_engineVersion :: Lens' RestoreDBClusterFromSnapshot (Maybe Text)

The version of the database engine to use for the new DB cluster.

restoreDBClusterFromSnapshot_deletionProtection :: Lens' RestoreDBClusterFromSnapshot (Maybe Bool)

A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

restoreDBClusterFromSnapshot_dbSubnetGroupName :: Lens' RestoreDBClusterFromSnapshot (Maybe Text)

The name of the DB subnet group to use for the new DB cluster.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

restoreDBClusterFromSnapshot_availabilityZones :: Lens' RestoreDBClusterFromSnapshot (Maybe [Text])

Provides the list of EC2 Availability Zones that instances in the restored DB cluster can be created in.

restoreDBClusterFromSnapshot_kmsKeyId :: Lens' RestoreDBClusterFromSnapshot (Maybe Text)

The Amazon KMS key identifier to use when restoring an encrypted DB cluster from a DB snapshot or DB cluster snapshot.

The KMS key identifier is the Amazon Resource Name (ARN) for the KMS encryption key. If you are restoring a DB cluster with the same Amazon account that owns the KMS encryption key used to encrypt the new DB cluster, then you can use the KMS key alias instead of the ARN for the KMS encryption key.

If you do not specify a value for the KmsKeyId parameter, then the following will occur:

If the DB snapshot or DB cluster snapshot in SnapshotIdentifier is encrypted, then the restored DB cluster is encrypted using the KMS key that was used to encrypt the DB snapshot or DB cluster snapshot.

If the DB snapshot or DB cluster snapshot in SnapshotIdentifier is not encrypted, then the restored DB cluster is not encrypted.

restoreDBClusterFromSnapshot_vpcSecurityGroupIds :: Lens' RestoreDBClusterFromSnapshot (Maybe [Text])

A list of VPC security groups that the new DB cluster will belong to.

restoreDBClusterFromSnapshot_databaseName :: Lens' RestoreDBClusterFromSnapshot (Maybe Text)

Not supported.

restoreDBClusterFromSnapshot_dbClusterParameterGroupName :: Lens' RestoreDBClusterFromSnapshot (Maybe Text)

The name of the DB cluster parameter group to associate with the new DB cluster.

Constraints:

If supplied, must match the name of an existing DBClusterParameterGroup.

restoreDBClusterFromSnapshot_optionGroupName :: Lens' RestoreDBClusterFromSnapshot (Maybe Text)

(Not supported by Neptune)

restoreDBClusterFromSnapshot_copyTagsToSnapshot :: Lens' RestoreDBClusterFromSnapshot (Maybe Bool)

/If set to true, tags are copied to any snapshot of the restored DB cluster that is created./

restoreDBClusterFromSnapshot_tags :: Lens' RestoreDBClusterFromSnapshot (Maybe [Tag])

The tags to be assigned to the restored DB cluster.

restoreDBClusterFromSnapshot_port :: Lens' RestoreDBClusterFromSnapshot (Maybe Int)

The port number on which the new DB cluster accepts connections.

Constraints: Value must be 1150-65535

Default: The same port as the original DB cluster.

restoreDBClusterFromSnapshot_enableIAMDatabaseAuthentication :: Lens' RestoreDBClusterFromSnapshot (Maybe Bool)

True to enable mapping of Amazon Identity and Access Management (IAM) accounts to database accounts, and otherwise false.

Default: false

restoreDBClusterFromSnapshot_enableCloudwatchLogsExports :: Lens' RestoreDBClusterFromSnapshot (Maybe [Text])

The list of logs that the restored DB cluster is to export to Amazon CloudWatch Logs.

restoreDBClusterFromSnapshot_dbClusterIdentifier :: Lens' RestoreDBClusterFromSnapshot Text

The name of the DB cluster to create from the DB snapshot or DB cluster snapshot. This parameter isn't case-sensitive.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

Example: my-snapshot-id

restoreDBClusterFromSnapshot_snapshotIdentifier :: Lens' RestoreDBClusterFromSnapshot Text

The identifier for the DB snapshot or DB cluster snapshot to restore from.

You can use either the name or the Amazon Resource Name (ARN) to specify a DB cluster snapshot. However, you can use only the ARN to specify a DB snapshot.

Constraints:

Must match the identifier of an existing Snapshot.

restoreDBClusterFromSnapshot_engine :: Lens' RestoreDBClusterFromSnapshot Text

The database engine to use for the new DB cluster.

Default: The same as source

Constraint: Must be compatible with the engine of the source

Destructuring the Response

data RestoreDBClusterFromSnapshotResponse

See: newRestoreDBClusterFromSnapshotResponse smart constructor.

Constructors

RestoreDBClusterFromSnapshotResponse'

Fields

dbCluster :: Maybe DBCluster

httpStatus :: Int

The response's http status code.

newRestoreDBClusterFromSnapshotResponse

Arguments

:: Int
$sel:httpStatus:RestoreDBClusterFromSnapshotResponse'

-> RestoreDBClusterFromSnapshotResponse

Create a value of RestoreDBClusterFromSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbCluster:RestoreDBClusterFromSnapshotResponse', restoreDBClusterFromSnapshotResponse_dbCluster - Undocumented member.

$sel:httpStatus:RestoreDBClusterFromSnapshotResponse', restoreDBClusterFromSnapshotResponse_httpStatus - The response's http status code.

Response Lenses

restoreDBClusterFromSnapshotResponse_dbCluster :: Lens' RestoreDBClusterFromSnapshotResponse (Maybe DBCluster)

Undocumented member.

restoreDBClusterFromSnapshotResponse_httpStatus :: Lens' RestoreDBClusterFromSnapshotResponse Int

The response's http status code.