Network.AWS.Neptune.Types.DBCluster

Description

Documentation

data DBCluster

Contains the details of an Amazon Neptune DB cluster.

This data type is used as a response element in the DescribeDBClusters action.

See: newDBCluster smart constructor.

Constructors

DBCluster'

Fields

engineVersion :: Maybe Text

Indicates the database engine version.

status :: Maybe Text

Specifies the current state of this DB cluster.

deletionProtection :: Maybe Bool

Indicates whether or not the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled.

automaticRestartTime :: Maybe ISO8601

Time at which the DB cluster will be automatically restarted.

storageEncrypted :: Maybe Bool

Specifies whether the DB cluster is encrypted.

dbClusterIdentifier :: Maybe Text

Contains a user-supplied DB cluster identifier. This identifier is the unique key that identifies a DB cluster.

dbClusterMembers :: Maybe [DBClusterMember]

Provides the list of instances that make up the DB cluster.

readReplicaIdentifiers :: Maybe [Text]

Contains one or more identifiers of the Read Replicas associated with this DB cluster.

replicationSourceIdentifier :: Maybe Text

Not supported by Neptune.

hostedZoneId :: Maybe Text

Specifies the ID that Amazon Route 53 assigns when you create a hosted zone.

dbClusterParameterGroup :: Maybe Text

Specifies the name of the DB cluster parameter group for the DB cluster.

masterUsername :: Maybe Text

Not supported by Neptune.

iAMDatabaseAuthenticationEnabled :: Maybe Bool

True if mapping of Amazon Identity and Access Management (IAM) accounts to database accounts is enabled, and otherwise false.

dbClusterResourceId :: Maybe Text

The Amazon Region-unique, immutable identifier for the DB cluster. This identifier is found in Amazon CloudTrail log entries whenever the Amazon KMS key for the DB cluster is accessed.

earliestRestorableTime :: Maybe ISO8601

Specifies the earliest time to which a database can be restored with point-in-time restore.

engine :: Maybe Text

Provides the name of the database engine to be used for this DB cluster.

dbClusterArn :: Maybe Text

The Amazon Resource Name (ARN) for the DB cluster.

cloneGroupId :: Maybe Text

Identifies the clone group to which the DB cluster is associated.

latestRestorableTime :: Maybe ISO8601

Specifies the latest time to which a database can be restored with point-in-time restore.

crossAccountClone :: Maybe Bool

If set to true, the DB cluster can be cloned across accounts.

preferredMaintenanceWindow :: Maybe Text

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

availabilityZones :: Maybe [Text]

Provides the list of EC2 Availability Zones that instances in the DB cluster can be created in.

characterSetName :: Maybe Text

Not supported by Neptune.

kmsKeyId :: Maybe Text

If StorageEncrypted is true, the Amazon KMS key identifier for the encrypted DB cluster.

preferredBackupWindow :: Maybe Text

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

associatedRoles :: Maybe [DBClusterRole]

Provides a list of the Amazon Identity and Access Management (IAM) roles that are associated with the DB cluster. IAM roles that are associated with a DB cluster grant permission for the DB cluster to access other Amazon services on your behalf.

vpcSecurityGroups :: Maybe [VpcSecurityGroupMembership]

Provides a list of VPC security groups that the DB cluster belongs to.

backupRetentionPeriod :: Maybe Int

Specifies the number of days for which automatic DB snapshots are retained.

dbSubnetGroup :: Maybe Text

Specifies information on the subnet group associated with the DB cluster, including the name, description, and subnets in the subnet group.

databaseName :: Maybe Text

Contains the name of the initial database of this DB cluster that was provided at create time, if one was specified when the DB cluster was created. This same name is returned for the life of the DB cluster.

multiAZ :: Maybe Bool

Specifies whether the DB cluster has instances in multiple Availability Zones.

enabledCloudwatchLogsExports :: Maybe [Text]

A list of log types that this DB cluster is configured to export to CloudWatch Logs.

allocatedStorage :: Maybe Int

AllocatedStorage always returns 1, because Neptune DB cluster storage size is not fixed, but instead automatically adjusts as needed.

copyTagsToSnapshot :: Maybe Bool

/If set to true, tags are copied to any snapshot of the DB cluster that is created./

clusterCreateTime :: Maybe ISO8601

Specifies the time when the DB cluster was created, in Universal Coordinated Time (UTC).

endpoint :: Maybe Text

Specifies the connection endpoint for the primary instance of the DB cluster.

percentProgress :: Maybe Text

Specifies the progress of the operation as a percentage.

readerEndpoint :: Maybe Text

The reader endpoint for the DB cluster. The reader endpoint for a DB cluster load-balances connections across the Read Replicas that are available in a DB cluster. As clients request new connections to the reader endpoint, Neptune distributes the connection requests among the Read Replicas in the DB cluster. This functionality can help balance your read workload across multiple Read Replicas in your DB cluster.

If a failover occurs, and the Read Replica that you are connected to is promoted to be the primary instance, your connection is dropped. To continue sending your read workload to other Read Replicas in the cluster, you can then reconnect to the reader endpoint.

port :: Maybe Int

Specifies the port that the database engine is listening on.

dbClusterOptionGroupMemberships :: Maybe [DBClusterOptionGroupStatus]

Not supported by Neptune.

newDBCluster :: DBCluster

Create a value of DBCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:DBCluster', dbCluster_engineVersion - Indicates the database engine version.

$sel:status:DBCluster', dbCluster_status - Specifies the current state of this DB cluster.

$sel:deletionProtection:DBCluster', dbCluster_deletionProtection - Indicates whether or not the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled.

$sel:automaticRestartTime:DBCluster', dbCluster_automaticRestartTime - Time at which the DB cluster will be automatically restarted.

$sel:storageEncrypted:DBCluster', dbCluster_storageEncrypted - Specifies whether the DB cluster is encrypted.

$sel:dbClusterIdentifier:DBCluster', dbCluster_dbClusterIdentifier - Contains a user-supplied DB cluster identifier. This identifier is the unique key that identifies a DB cluster.

$sel:dbClusterMembers:DBCluster', dbCluster_dbClusterMembers - Provides the list of instances that make up the DB cluster.

$sel:readReplicaIdentifiers:DBCluster', dbCluster_readReplicaIdentifiers - Contains one or more identifiers of the Read Replicas associated with this DB cluster.

$sel:replicationSourceIdentifier:DBCluster', dbCluster_replicationSourceIdentifier - Not supported by Neptune.

$sel:hostedZoneId:DBCluster', dbCluster_hostedZoneId - Specifies the ID that Amazon Route 53 assigns when you create a hosted zone.

$sel:dbClusterParameterGroup:DBCluster', dbCluster_dbClusterParameterGroup - Specifies the name of the DB cluster parameter group for the DB cluster.

$sel:masterUsername:DBCluster', dbCluster_masterUsername - Not supported by Neptune.

$sel:iAMDatabaseAuthenticationEnabled:DBCluster', dbCluster_iAMDatabaseAuthenticationEnabled - True if mapping of Amazon Identity and Access Management (IAM) accounts to database accounts is enabled, and otherwise false.

$sel:dbClusterResourceId:DBCluster', dbCluster_dbClusterResourceId - The Amazon Region-unique, immutable identifier for the DB cluster. This identifier is found in Amazon CloudTrail log entries whenever the Amazon KMS key for the DB cluster is accessed.

$sel:earliestRestorableTime:DBCluster', dbCluster_earliestRestorableTime - Specifies the earliest time to which a database can be restored with point-in-time restore.

$sel:engine:DBCluster', dbCluster_engine - Provides the name of the database engine to be used for this DB cluster.

$sel:dbClusterArn:DBCluster', dbCluster_dbClusterArn - The Amazon Resource Name (ARN) for the DB cluster.

$sel:cloneGroupId:DBCluster', dbCluster_cloneGroupId - Identifies the clone group to which the DB cluster is associated.

$sel:latestRestorableTime:DBCluster', dbCluster_latestRestorableTime - Specifies the latest time to which a database can be restored with point-in-time restore.

$sel:crossAccountClone:DBCluster', dbCluster_crossAccountClone - If set to true, the DB cluster can be cloned across accounts.

$sel:preferredMaintenanceWindow:DBCluster', dbCluster_preferredMaintenanceWindow - Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

$sel:availabilityZones:DBCluster', dbCluster_availabilityZones - Provides the list of EC2 Availability Zones that instances in the DB cluster can be created in.

$sel:characterSetName:DBCluster', dbCluster_characterSetName - Not supported by Neptune.

$sel:kmsKeyId:DBCluster', dbCluster_kmsKeyId - If StorageEncrypted is true, the Amazon KMS key identifier for the encrypted DB cluster.

$sel:preferredBackupWindow:DBCluster', dbCluster_preferredBackupWindow - Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

$sel:associatedRoles:DBCluster', dbCluster_associatedRoles - Provides a list of the Amazon Identity and Access Management (IAM) roles that are associated with the DB cluster. IAM roles that are associated with a DB cluster grant permission for the DB cluster to access other Amazon services on your behalf.

$sel:vpcSecurityGroups:DBCluster', dbCluster_vpcSecurityGroups - Provides a list of VPC security groups that the DB cluster belongs to.

$sel:backupRetentionPeriod:DBCluster', dbCluster_backupRetentionPeriod - Specifies the number of days for which automatic DB snapshots are retained.

$sel:dbSubnetGroup:DBCluster', dbCluster_dbSubnetGroup - Specifies information on the subnet group associated with the DB cluster, including the name, description, and subnets in the subnet group.

$sel:databaseName:DBCluster', dbCluster_databaseName - Contains the name of the initial database of this DB cluster that was provided at create time, if one was specified when the DB cluster was created. This same name is returned for the life of the DB cluster.

$sel:multiAZ:DBCluster', dbCluster_multiAZ - Specifies whether the DB cluster has instances in multiple Availability Zones.

$sel:enabledCloudwatchLogsExports:DBCluster', dbCluster_enabledCloudwatchLogsExports - A list of log types that this DB cluster is configured to export to CloudWatch Logs.

$sel:allocatedStorage:DBCluster', dbCluster_allocatedStorage - AllocatedStorage always returns 1, because Neptune DB cluster storage size is not fixed, but instead automatically adjusts as needed.

$sel:copyTagsToSnapshot:DBCluster', dbCluster_copyTagsToSnapshot - /If set to true, tags are copied to any snapshot of the DB cluster that is created./

$sel:clusterCreateTime:DBCluster', dbCluster_clusterCreateTime - Specifies the time when the DB cluster was created, in Universal Coordinated Time (UTC).

$sel:endpoint:DBCluster', dbCluster_endpoint - Specifies the connection endpoint for the primary instance of the DB cluster.

$sel:percentProgress:DBCluster', dbCluster_percentProgress - Specifies the progress of the operation as a percentage.

$sel:readerEndpoint:DBCluster', dbCluster_readerEndpoint - The reader endpoint for the DB cluster. The reader endpoint for a DB cluster load-balances connections across the Read Replicas that are available in a DB cluster. As clients request new connections to the reader endpoint, Neptune distributes the connection requests among the Read Replicas in the DB cluster. This functionality can help balance your read workload across multiple Read Replicas in your DB cluster.

If a failover occurs, and the Read Replica that you are connected to is promoted to be the primary instance, your connection is dropped. To continue sending your read workload to other Read Replicas in the cluster, you can then reconnect to the reader endpoint.

$sel:port:DBCluster', dbCluster_port - Specifies the port that the database engine is listening on.

$sel:dbClusterOptionGroupMemberships:DBCluster', dbCluster_dbClusterOptionGroupMemberships - Not supported by Neptune.

dbCluster_engineVersion :: Lens' DBCluster (Maybe Text)

Indicates the database engine version.

dbCluster_status :: Lens' DBCluster (Maybe Text)

Specifies the current state of this DB cluster.

dbCluster_deletionProtection :: Lens' DBCluster (Maybe Bool)

Indicates whether or not the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled.

dbCluster_automaticRestartTime :: Lens' DBCluster (Maybe UTCTime)

Time at which the DB cluster will be automatically restarted.

dbCluster_storageEncrypted :: Lens' DBCluster (Maybe Bool)

Specifies whether the DB cluster is encrypted.

dbCluster_dbClusterIdentifier :: Lens' DBCluster (Maybe Text)

Contains a user-supplied DB cluster identifier. This identifier is the unique key that identifies a DB cluster.

dbCluster_dbClusterMembers :: Lens' DBCluster (Maybe [DBClusterMember])

Provides the list of instances that make up the DB cluster.

dbCluster_readReplicaIdentifiers :: Lens' DBCluster (Maybe [Text])

Contains one or more identifiers of the Read Replicas associated with this DB cluster.

dbCluster_replicationSourceIdentifier :: Lens' DBCluster (Maybe Text)

Not supported by Neptune.

dbCluster_hostedZoneId :: Lens' DBCluster (Maybe Text)

Specifies the ID that Amazon Route 53 assigns when you create a hosted zone.

dbCluster_dbClusterParameterGroup :: Lens' DBCluster (Maybe Text)

Specifies the name of the DB cluster parameter group for the DB cluster.

dbCluster_masterUsername :: Lens' DBCluster (Maybe Text)

Not supported by Neptune.

dbCluster_iAMDatabaseAuthenticationEnabled :: Lens' DBCluster (Maybe Bool)

True if mapping of Amazon Identity and Access Management (IAM) accounts to database accounts is enabled, and otherwise false.

dbCluster_dbClusterResourceId :: Lens' DBCluster (Maybe Text)

The Amazon Region-unique, immutable identifier for the DB cluster. This identifier is found in Amazon CloudTrail log entries whenever the Amazon KMS key for the DB cluster is accessed.

dbCluster_earliestRestorableTime :: Lens' DBCluster (Maybe UTCTime)

Specifies the earliest time to which a database can be restored with point-in-time restore.

dbCluster_engine :: Lens' DBCluster (Maybe Text)

Provides the name of the database engine to be used for this DB cluster.

dbCluster_dbClusterArn :: Lens' DBCluster (Maybe Text)

The Amazon Resource Name (ARN) for the DB cluster.

dbCluster_cloneGroupId :: Lens' DBCluster (Maybe Text)

Identifies the clone group to which the DB cluster is associated.

dbCluster_latestRestorableTime :: Lens' DBCluster (Maybe UTCTime)

Specifies the latest time to which a database can be restored with point-in-time restore.

dbCluster_crossAccountClone :: Lens' DBCluster (Maybe Bool)

If set to true, the DB cluster can be cloned across accounts.

dbCluster_preferredMaintenanceWindow :: Lens' DBCluster (Maybe Text)

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

dbCluster_availabilityZones :: Lens' DBCluster (Maybe [Text])

Provides the list of EC2 Availability Zones that instances in the DB cluster can be created in.

dbCluster_characterSetName :: Lens' DBCluster (Maybe Text)

Not supported by Neptune.

dbCluster_kmsKeyId :: Lens' DBCluster (Maybe Text)

If StorageEncrypted is true, the Amazon KMS key identifier for the encrypted DB cluster.

dbCluster_preferredBackupWindow :: Lens' DBCluster (Maybe Text)

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

dbCluster_associatedRoles :: Lens' DBCluster (Maybe [DBClusterRole])

Provides a list of the Amazon Identity and Access Management (IAM) roles that are associated with the DB cluster. IAM roles that are associated with a DB cluster grant permission for the DB cluster to access other Amazon services on your behalf.

dbCluster_vpcSecurityGroups :: Lens' DBCluster (Maybe [VpcSecurityGroupMembership])

Provides a list of VPC security groups that the DB cluster belongs to.

dbCluster_backupRetentionPeriod :: Lens' DBCluster (Maybe Int)

Specifies the number of days for which automatic DB snapshots are retained.

dbCluster_dbSubnetGroup :: Lens' DBCluster (Maybe Text)

Specifies information on the subnet group associated with the DB cluster, including the name, description, and subnets in the subnet group.

dbCluster_databaseName :: Lens' DBCluster (Maybe Text)

Contains the name of the initial database of this DB cluster that was provided at create time, if one was specified when the DB cluster was created. This same name is returned for the life of the DB cluster.

dbCluster_multiAZ :: Lens' DBCluster (Maybe Bool)

Specifies whether the DB cluster has instances in multiple Availability Zones.

dbCluster_enabledCloudwatchLogsExports :: Lens' DBCluster (Maybe [Text])

A list of log types that this DB cluster is configured to export to CloudWatch Logs.

dbCluster_allocatedStorage :: Lens' DBCluster (Maybe Int)

AllocatedStorage always returns 1, because Neptune DB cluster storage size is not fixed, but instead automatically adjusts as needed.

dbCluster_copyTagsToSnapshot :: Lens' DBCluster (Maybe Bool)

/If set to true, tags are copied to any snapshot of the DB cluster that is created./

dbCluster_clusterCreateTime :: Lens' DBCluster (Maybe UTCTime)

Specifies the time when the DB cluster was created, in Universal Coordinated Time (UTC).

dbCluster_endpoint :: Lens' DBCluster (Maybe Text)

Specifies the connection endpoint for the primary instance of the DB cluster.

dbCluster_percentProgress :: Lens' DBCluster (Maybe Text)

Specifies the progress of the operation as a percentage.

dbCluster_readerEndpoint :: Lens' DBCluster (Maybe Text)

The reader endpoint for the DB cluster. The reader endpoint for a DB cluster load-balances connections across the Read Replicas that are available in a DB cluster. As clients request new connections to the reader endpoint, Neptune distributes the connection requests among the Read Replicas in the DB cluster. This functionality can help balance your read workload across multiple Read Replicas in your DB cluster.

If a failover occurs, and the Read Replica that you are connected to is promoted to be the primary instance, your connection is dropped. To continue sending your read workload to other Read Replicas in the cluster, you can then reconnect to the reader endpoint.

dbCluster_port :: Lens' DBCluster (Maybe Int)

Specifies the port that the database engine is listening on.

dbCluster_dbClusterOptionGroupMemberships :: Lens' DBCluster (Maybe [DBClusterOptionGroupStatus])

Not supported by Neptune.