Network.AWS.Neptune.Types.DBInstance

Description

Documentation

data DBInstance

Contains the details of an Amazon Neptune DB instance.

This data type is used as a response element in the DescribeDBInstances action.

See: newDBInstance smart constructor.

Constructors

DBInstance'

Fields

engineVersion :: Maybe Text

Indicates the database engine version.

dbSecurityGroups :: Maybe [DBSecurityGroupMembership]

Provides List of DB security group elements containing only DBSecurityGroup.Name and DBSecurityGroup.Status subelements.

deletionProtection :: Maybe Bool

Indicates whether or not the DB instance has deletion protection enabled. The instance can't be deleted when deletion protection is enabled. See Deleting a DB Instance.

storageEncrypted :: Maybe Bool

Not supported: The encryption for DB instances is managed by the DB cluster.

dbClusterIdentifier :: Maybe Text

If the DB instance is a member of a DB cluster, contains the name of the DB cluster that the DB instance is a member of.

publiclyAccessible :: Maybe Bool

This flag should no longer be used.

autoMinorVersionUpgrade :: Maybe Bool

Indicates that minor version patches are applied automatically.

dbInstanceArn :: Maybe Text

The Amazon Resource Name (ARN) for the DB instance.

masterUsername :: Maybe Text

Not supported by Neptune.

readReplicaDBInstanceIdentifiers :: Maybe [Text]

Contains one or more identifiers of the Read Replicas associated with this DB instance.

iAMDatabaseAuthenticationEnabled :: Maybe Bool

True if Amazon Identity and Access Management (IAM) authentication is enabled, and otherwise false.

monitoringRoleArn :: Maybe Text

The ARN for the IAM role that permits Neptune to send Enhanced Monitoring metrics to Amazon CloudWatch Logs.

iops :: Maybe Int

Specifies the Provisioned IOPS (I/O operations per second) value.

instanceCreateTime :: Maybe ISO8601

Provides the date and time the DB instance was created.

readReplicaSourceDBInstanceIdentifier :: Maybe Text

Contains the identifier of the source DB instance if this DB instance is a Read Replica.

monitoringInterval :: Maybe Int

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance.

engine :: Maybe Text

Provides the name of the database engine to be used for this DB instance.

latestRestorableTime :: Maybe ISO8601

Specifies the latest time to which a database can be restored with point-in-time restore.

dbInstanceClass :: Maybe Text

Contains the name of the compute and memory capacity class of the DB instance.

promotionTier :: Maybe Int

A value that specifies the order in which a Read Replica is promoted to the primary instance after a failure of the existing primary instance.

licenseModel :: Maybe Text

License model information for this DB instance.

preferredMaintenanceWindow :: Maybe Text

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

cACertificateIdentifier :: Maybe Text

The identifier of the CA certificate for this DB instance.

dbInstanceIdentifier :: Maybe Text

Contains a user-supplied database identifier. This identifier is the unique key that identifies a DB instance.

characterSetName :: Maybe Text

(Not supported by Neptune)

kmsKeyId :: Maybe Text

Not supported: The encryption for DB instances is managed by the DB cluster.

preferredBackupWindow :: Maybe Text

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

availabilityZone :: Maybe Text

Specifies the name of the Availability Zone the DB instance is located in.

vpcSecurityGroups :: Maybe [VpcSecurityGroupMembership]

Provides a list of VPC security group elements that the DB instance belongs to.

backupRetentionPeriod :: Maybe Int

Specifies the number of days for which automatic DB snapshots are retained.

performanceInsightsKMSKeyId :: Maybe Text

(Not supported by Neptune)

dbSubnetGroup :: Maybe DBSubnetGroup

Specifies information on the subnet group associated with the DB instance, including the name, description, and subnets in the subnet group.

multiAZ :: Maybe Bool

Specifies if the DB instance is a Multi-AZ deployment.

optionGroupMemberships :: Maybe [OptionGroupMembership]

(Not supported by Neptune)

enabledCloudwatchLogsExports :: Maybe [Text]

A list of log types that this DB instance is configured to export to CloudWatch Logs.

enhancedMonitoringResourceArn :: Maybe Text

The Amazon Resource Name (ARN) of the Amazon CloudWatch Logs log stream that receives the Enhanced Monitoring metrics data for the DB instance.

secondaryAvailabilityZone :: Maybe Text

If present, specifies the name of the secondary Availability Zone for a DB instance with multi-AZ support.

performanceInsightsEnabled :: Maybe Bool

(Not supported by Neptune)

allocatedStorage :: Maybe Int

Not supported by Neptune.

dbiResourceId :: Maybe Text

The Amazon Region-unique, immutable identifier for the DB instance. This identifier is found in Amazon CloudTrail log entries whenever the Amazon KMS key for the DB instance is accessed.

dbParameterGroups :: Maybe [DBParameterGroupStatus]

Provides the list of DB parameter groups applied to this DB instance.

copyTagsToSnapshot :: Maybe Bool

Specifies whether tags are copied from the DB instance to snapshots of the DB instance.

timezone :: Maybe Text

Not supported.

tdeCredentialArn :: Maybe Text

The ARN from the key store with which the instance is associated for TDE encryption.

endpoint :: Maybe Endpoint

Specifies the connection endpoint.

dbInstanceStatus :: Maybe Text

Specifies the current state of this database.

dbInstancePort :: Maybe Int

Specifies the port that the DB instance listens on. If the DB instance is part of a DB cluster, this can be a different port than the DB cluster port.

pendingModifiedValues :: Maybe PendingModifiedValues

Specifies that changes to the DB instance are pending. This element is only included when changes are pending. Specific changes are identified by subelements.

readReplicaDBClusterIdentifiers :: Maybe [Text]

Contains one or more identifiers of DB clusters that are Read Replicas of this DB instance.

storageType :: Maybe Text

Specifies the storage type associated with DB instance.

statusInfos :: Maybe [DBInstanceStatusInfo]

The status of a Read Replica. If the instance is not a Read Replica, this is blank.

domainMemberships :: Maybe [DomainMembership]

Not supported

dbName :: Maybe Text

The database name.

newDBInstance :: DBInstance

Create a value of DBInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:DBInstance', dbInstance_engineVersion - Indicates the database engine version.

$sel:dbSecurityGroups:DBInstance', dbInstance_dbSecurityGroups - Provides List of DB security group elements containing only DBSecurityGroup.Name and DBSecurityGroup.Status subelements.

$sel:deletionProtection:DBInstance', dbInstance_deletionProtection - Indicates whether or not the DB instance has deletion protection enabled. The instance can't be deleted when deletion protection is enabled. See Deleting a DB Instance.

$sel:storageEncrypted:DBInstance', dbInstance_storageEncrypted - Not supported: The encryption for DB instances is managed by the DB cluster.

$sel:dbClusterIdentifier:DBInstance', dbInstance_dbClusterIdentifier - If the DB instance is a member of a DB cluster, contains the name of the DB cluster that the DB instance is a member of.

$sel:publiclyAccessible:DBInstance', dbInstance_publiclyAccessible - This flag should no longer be used.

$sel:autoMinorVersionUpgrade:DBInstance', dbInstance_autoMinorVersionUpgrade - Indicates that minor version patches are applied automatically.

$sel:dbInstanceArn:DBInstance', dbInstance_dbInstanceArn - The Amazon Resource Name (ARN) for the DB instance.

$sel:masterUsername:DBInstance', dbInstance_masterUsername - Not supported by Neptune.

$sel:readReplicaDBInstanceIdentifiers:DBInstance', dbInstance_readReplicaDBInstanceIdentifiers - Contains one or more identifiers of the Read Replicas associated with this DB instance.

$sel:iAMDatabaseAuthenticationEnabled:DBInstance', dbInstance_iAMDatabaseAuthenticationEnabled - True if Amazon Identity and Access Management (IAM) authentication is enabled, and otherwise false.

$sel:monitoringRoleArn:DBInstance', dbInstance_monitoringRoleArn - The ARN for the IAM role that permits Neptune to send Enhanced Monitoring metrics to Amazon CloudWatch Logs.

$sel:iops:DBInstance', dbInstance_iops - Specifies the Provisioned IOPS (I/O operations per second) value.

$sel:instanceCreateTime:DBInstance', dbInstance_instanceCreateTime - Provides the date and time the DB instance was created.

$sel:readReplicaSourceDBInstanceIdentifier:DBInstance', dbInstance_readReplicaSourceDBInstanceIdentifier - Contains the identifier of the source DB instance if this DB instance is a Read Replica.

$sel:monitoringInterval:DBInstance', dbInstance_monitoringInterval - The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance.

$sel:engine:DBInstance', dbInstance_engine - Provides the name of the database engine to be used for this DB instance.

$sel:latestRestorableTime:DBInstance', dbInstance_latestRestorableTime - Specifies the latest time to which a database can be restored with point-in-time restore.

$sel:dbInstanceClass:DBInstance', dbInstance_dbInstanceClass - Contains the name of the compute and memory capacity class of the DB instance.

$sel:promotionTier:DBInstance', dbInstance_promotionTier - A value that specifies the order in which a Read Replica is promoted to the primary instance after a failure of the existing primary instance.

$sel:licenseModel:DBInstance', dbInstance_licenseModel - License model information for this DB instance.

$sel:preferredMaintenanceWindow:DBInstance', dbInstance_preferredMaintenanceWindow - Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

$sel:cACertificateIdentifier:DBInstance', dbInstance_cACertificateIdentifier - The identifier of the CA certificate for this DB instance.

$sel:dbInstanceIdentifier:DBInstance', dbInstance_dbInstanceIdentifier - Contains a user-supplied database identifier. This identifier is the unique key that identifies a DB instance.

$sel:characterSetName:DBInstance', dbInstance_characterSetName - (Not supported by Neptune)

$sel:kmsKeyId:DBInstance', dbInstance_kmsKeyId - Not supported: The encryption for DB instances is managed by the DB cluster.

$sel:preferredBackupWindow:DBInstance', dbInstance_preferredBackupWindow - Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

$sel:availabilityZone:DBInstance', dbInstance_availabilityZone - Specifies the name of the Availability Zone the DB instance is located in.

$sel:vpcSecurityGroups:DBInstance', dbInstance_vpcSecurityGroups - Provides a list of VPC security group elements that the DB instance belongs to.

$sel:backupRetentionPeriod:DBInstance', dbInstance_backupRetentionPeriod - Specifies the number of days for which automatic DB snapshots are retained.

$sel:performanceInsightsKMSKeyId:DBInstance', dbInstance_performanceInsightsKMSKeyId - (Not supported by Neptune)

$sel:dbSubnetGroup:DBInstance', dbInstance_dbSubnetGroup - Specifies information on the subnet group associated with the DB instance, including the name, description, and subnets in the subnet group.

$sel:multiAZ:DBInstance', dbInstance_multiAZ - Specifies if the DB instance is a Multi-AZ deployment.

$sel:optionGroupMemberships:DBInstance', dbInstance_optionGroupMemberships - (Not supported by Neptune)

$sel:enabledCloudwatchLogsExports:DBInstance', dbInstance_enabledCloudwatchLogsExports - A list of log types that this DB instance is configured to export to CloudWatch Logs.

$sel:enhancedMonitoringResourceArn:DBInstance', dbInstance_enhancedMonitoringResourceArn - The Amazon Resource Name (ARN) of the Amazon CloudWatch Logs log stream that receives the Enhanced Monitoring metrics data for the DB instance.

$sel:secondaryAvailabilityZone:DBInstance', dbInstance_secondaryAvailabilityZone - If present, specifies the name of the secondary Availability Zone for a DB instance with multi-AZ support.

$sel:performanceInsightsEnabled:DBInstance', dbInstance_performanceInsightsEnabled - (Not supported by Neptune)

$sel:allocatedStorage:DBInstance', dbInstance_allocatedStorage - Not supported by Neptune.

$sel:dbiResourceId:DBInstance', dbInstance_dbiResourceId - The Amazon Region-unique, immutable identifier for the DB instance. This identifier is found in Amazon CloudTrail log entries whenever the Amazon KMS key for the DB instance is accessed.

$sel:dbParameterGroups:DBInstance', dbInstance_dbParameterGroups - Provides the list of DB parameter groups applied to this DB instance.

$sel:copyTagsToSnapshot:DBInstance', dbInstance_copyTagsToSnapshot - Specifies whether tags are copied from the DB instance to snapshots of the DB instance.

$sel:timezone:DBInstance', dbInstance_timezone - Not supported.

$sel:tdeCredentialArn:DBInstance', dbInstance_tdeCredentialArn - The ARN from the key store with which the instance is associated for TDE encryption.

$sel:endpoint:DBInstance', dbInstance_endpoint - Specifies the connection endpoint.

$sel:dbInstanceStatus:DBInstance', dbInstance_dbInstanceStatus - Specifies the current state of this database.

$sel:dbInstancePort:DBInstance', dbInstance_dbInstancePort - Specifies the port that the DB instance listens on. If the DB instance is part of a DB cluster, this can be a different port than the DB cluster port.

$sel:pendingModifiedValues:DBInstance', dbInstance_pendingModifiedValues - Specifies that changes to the DB instance are pending. This element is only included when changes are pending. Specific changes are identified by subelements.

$sel:readReplicaDBClusterIdentifiers:DBInstance', dbInstance_readReplicaDBClusterIdentifiers - Contains one or more identifiers of DB clusters that are Read Replicas of this DB instance.

$sel:storageType:DBInstance', dbInstance_storageType - Specifies the storage type associated with DB instance.

$sel:statusInfos:DBInstance', dbInstance_statusInfos - The status of a Read Replica. If the instance is not a Read Replica, this is blank.

$sel:domainMemberships:DBInstance', dbInstance_domainMemberships - Not supported

$sel:dbName:DBInstance', dbInstance_dbName - The database name.

dbInstance_engineVersion :: Lens' DBInstance (Maybe Text)

Indicates the database engine version.

dbInstance_dbSecurityGroups :: Lens' DBInstance (Maybe [DBSecurityGroupMembership])

Provides List of DB security group elements containing only DBSecurityGroup.Name and DBSecurityGroup.Status subelements.

dbInstance_deletionProtection :: Lens' DBInstance (Maybe Bool)

Indicates whether or not the DB instance has deletion protection enabled. The instance can't be deleted when deletion protection is enabled. See Deleting a DB Instance.

dbInstance_storageEncrypted :: Lens' DBInstance (Maybe Bool)

Not supported: The encryption for DB instances is managed by the DB cluster.

dbInstance_dbClusterIdentifier :: Lens' DBInstance (Maybe Text)

If the DB instance is a member of a DB cluster, contains the name of the DB cluster that the DB instance is a member of.

dbInstance_publiclyAccessible :: Lens' DBInstance (Maybe Bool)

This flag should no longer be used.

dbInstance_autoMinorVersionUpgrade :: Lens' DBInstance (Maybe Bool)

Indicates that minor version patches are applied automatically.

dbInstance_dbInstanceArn :: Lens' DBInstance (Maybe Text)

The Amazon Resource Name (ARN) for the DB instance.

dbInstance_masterUsername :: Lens' DBInstance (Maybe Text)

Not supported by Neptune.

dbInstance_readReplicaDBInstanceIdentifiers :: Lens' DBInstance (Maybe [Text])

Contains one or more identifiers of the Read Replicas associated with this DB instance.

dbInstance_iAMDatabaseAuthenticationEnabled :: Lens' DBInstance (Maybe Bool)

True if Amazon Identity and Access Management (IAM) authentication is enabled, and otherwise false.

dbInstance_monitoringRoleArn :: Lens' DBInstance (Maybe Text)

The ARN for the IAM role that permits Neptune to send Enhanced Monitoring metrics to Amazon CloudWatch Logs.

dbInstance_iops :: Lens' DBInstance (Maybe Int)

Specifies the Provisioned IOPS (I/O operations per second) value.

dbInstance_instanceCreateTime :: Lens' DBInstance (Maybe UTCTime)

Provides the date and time the DB instance was created.

dbInstance_readReplicaSourceDBInstanceIdentifier :: Lens' DBInstance (Maybe Text)

Contains the identifier of the source DB instance if this DB instance is a Read Replica.

dbInstance_monitoringInterval :: Lens' DBInstance (Maybe Int)

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance.

dbInstance_engine :: Lens' DBInstance (Maybe Text)

Provides the name of the database engine to be used for this DB instance.

dbInstance_latestRestorableTime :: Lens' DBInstance (Maybe UTCTime)

Specifies the latest time to which a database can be restored with point-in-time restore.

dbInstance_dbInstanceClass :: Lens' DBInstance (Maybe Text)

Contains the name of the compute and memory capacity class of the DB instance.

dbInstance_promotionTier :: Lens' DBInstance (Maybe Int)

A value that specifies the order in which a Read Replica is promoted to the primary instance after a failure of the existing primary instance.

dbInstance_licenseModel :: Lens' DBInstance (Maybe Text)

License model information for this DB instance.

dbInstance_preferredMaintenanceWindow :: Lens' DBInstance (Maybe Text)

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

dbInstance_cACertificateIdentifier :: Lens' DBInstance (Maybe Text)

The identifier of the CA certificate for this DB instance.

dbInstance_dbInstanceIdentifier :: Lens' DBInstance (Maybe Text)

Contains a user-supplied database identifier. This identifier is the unique key that identifies a DB instance.

dbInstance_characterSetName :: Lens' DBInstance (Maybe Text)

(Not supported by Neptune)

dbInstance_kmsKeyId :: Lens' DBInstance (Maybe Text)

Not supported: The encryption for DB instances is managed by the DB cluster.

dbInstance_preferredBackupWindow :: Lens' DBInstance (Maybe Text)

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

dbInstance_availabilityZone :: Lens' DBInstance (Maybe Text)

Specifies the name of the Availability Zone the DB instance is located in.

dbInstance_vpcSecurityGroups :: Lens' DBInstance (Maybe [VpcSecurityGroupMembership])

Provides a list of VPC security group elements that the DB instance belongs to.

dbInstance_backupRetentionPeriod :: Lens' DBInstance (Maybe Int)

Specifies the number of days for which automatic DB snapshots are retained.

dbInstance_performanceInsightsKMSKeyId :: Lens' DBInstance (Maybe Text)

(Not supported by Neptune)

dbInstance_dbSubnetGroup :: Lens' DBInstance (Maybe DBSubnetGroup)

Specifies information on the subnet group associated with the DB instance, including the name, description, and subnets in the subnet group.

dbInstance_multiAZ :: Lens' DBInstance (Maybe Bool)

Specifies if the DB instance is a Multi-AZ deployment.

dbInstance_optionGroupMemberships :: Lens' DBInstance (Maybe [OptionGroupMembership])

(Not supported by Neptune)

dbInstance_enabledCloudwatchLogsExports :: Lens' DBInstance (Maybe [Text])

A list of log types that this DB instance is configured to export to CloudWatch Logs.

dbInstance_enhancedMonitoringResourceArn :: Lens' DBInstance (Maybe Text)

The Amazon Resource Name (ARN) of the Amazon CloudWatch Logs log stream that receives the Enhanced Monitoring metrics data for the DB instance.

dbInstance_secondaryAvailabilityZone :: Lens' DBInstance (Maybe Text)

If present, specifies the name of the secondary Availability Zone for a DB instance with multi-AZ support.

dbInstance_performanceInsightsEnabled :: Lens' DBInstance (Maybe Bool)

(Not supported by Neptune)

dbInstance_allocatedStorage :: Lens' DBInstance (Maybe Int)

Not supported by Neptune.

dbInstance_dbiResourceId :: Lens' DBInstance (Maybe Text)

The Amazon Region-unique, immutable identifier for the DB instance. This identifier is found in Amazon CloudTrail log entries whenever the Amazon KMS key for the DB instance is accessed.

dbInstance_dbParameterGroups :: Lens' DBInstance (Maybe [DBParameterGroupStatus])

Provides the list of DB parameter groups applied to this DB instance.

dbInstance_copyTagsToSnapshot :: Lens' DBInstance (Maybe Bool)

Specifies whether tags are copied from the DB instance to snapshots of the DB instance.

dbInstance_timezone :: Lens' DBInstance (Maybe Text)

Not supported.

dbInstance_tdeCredentialArn :: Lens' DBInstance (Maybe Text)

The ARN from the key store with which the instance is associated for TDE encryption.

dbInstance_endpoint :: Lens' DBInstance (Maybe Endpoint)

Specifies the connection endpoint.

dbInstance_dbInstanceStatus :: Lens' DBInstance (Maybe Text)

Specifies the current state of this database.

dbInstance_dbInstancePort :: Lens' DBInstance (Maybe Int)

Specifies the port that the DB instance listens on. If the DB instance is part of a DB cluster, this can be a different port than the DB cluster port.

dbInstance_pendingModifiedValues :: Lens' DBInstance (Maybe PendingModifiedValues)

Specifies that changes to the DB instance are pending. This element is only included when changes are pending. Specific changes are identified by subelements.

dbInstance_readReplicaDBClusterIdentifiers :: Lens' DBInstance (Maybe [Text])

Contains one or more identifiers of DB clusters that are Read Replicas of this DB instance.

dbInstance_storageType :: Lens' DBInstance (Maybe Text)

Specifies the storage type associated with DB instance.

dbInstance_statusInfos :: Lens' DBInstance (Maybe [DBInstanceStatusInfo])

The status of a Read Replica. If the instance is not a Read Replica, this is blank.

dbInstance_domainMemberships :: Lens' DBInstance (Maybe [DomainMembership])

Not supported

dbInstance_dbName :: Lens' DBInstance (Maybe Text)

The database name.