Network.AWS.NetworkFirewall.AssociateFirewallPolicy

Description

Associates a FirewallPolicy to a Firewall.

A firewall policy defines how to monitor and manage your VPC network traffic, using a collection of inspection rule groups and other settings. Each firewall requires one firewall policy association, and you can use the same firewall policy for multiple firewalls.

Creating a Request

data AssociateFirewallPolicy

See: newAssociateFirewallPolicy smart constructor.

Constructors

AssociateFirewallPolicy'

Fields

updateToken :: Maybe Text

An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

firewallArn :: Maybe Text

The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.

firewallName :: Maybe Text

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

You must specify the ARN or the name, and you can specify both.

firewallPolicyArn :: Text

The Amazon Resource Name (ARN) of the firewall policy.

newAssociateFirewallPolicy

Arguments

:: Text
$sel:firewallPolicyArn:AssociateFirewallPolicy'

-> AssociateFirewallPolicy

Create a value of AssociateFirewallPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateToken:AssociateFirewallPolicy', associateFirewallPolicy_updateToken - An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

$sel:firewallArn:AssociateFirewallPolicy', associateFirewallPolicy_firewallArn - The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.

$sel:firewallName:AssociateFirewallPolicy', associateFirewallPolicy_firewallName - The descriptive name of the firewall. You can't change the name of a firewall after you create it.

You must specify the ARN or the name, and you can specify both.

$sel:firewallPolicyArn:AssociateFirewallPolicy', associateFirewallPolicy_firewallPolicyArn - The Amazon Resource Name (ARN) of the firewall policy.

Request Lenses

associateFirewallPolicy_updateToken :: Lens' AssociateFirewallPolicy (Maybe Text)

An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

associateFirewallPolicy_firewallArn :: Lens' AssociateFirewallPolicy (Maybe Text)

The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.

associateFirewallPolicy_firewallName :: Lens' AssociateFirewallPolicy (Maybe Text)

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

You must specify the ARN or the name, and you can specify both.

associateFirewallPolicy_firewallPolicyArn :: Lens' AssociateFirewallPolicy Text

The Amazon Resource Name (ARN) of the firewall policy.

Destructuring the Response

data AssociateFirewallPolicyResponse

See: newAssociateFirewallPolicyResponse smart constructor.

Constructors

AssociateFirewallPolicyResponse'

Fields

updateToken :: Maybe Text

An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

firewallArn :: Maybe Text

The Amazon Resource Name (ARN) of the firewall.

firewallPolicyArn :: Maybe Text

The Amazon Resource Name (ARN) of the firewall policy.

firewallName :: Maybe Text

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

httpStatus :: Int

The response's http status code.

newAssociateFirewallPolicyResponse

Arguments

:: Int
$sel:httpStatus:AssociateFirewallPolicyResponse'

-> AssociateFirewallPolicyResponse

Create a value of AssociateFirewallPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateToken:AssociateFirewallPolicy', associateFirewallPolicyResponse_updateToken - An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

$sel:firewallArn:AssociateFirewallPolicy', associateFirewallPolicyResponse_firewallArn - The Amazon Resource Name (ARN) of the firewall.

$sel:firewallPolicyArn:AssociateFirewallPolicy', associateFirewallPolicyResponse_firewallPolicyArn - The Amazon Resource Name (ARN) of the firewall policy.

$sel:firewallName:AssociateFirewallPolicy', associateFirewallPolicyResponse_firewallName - The descriptive name of the firewall. You can't change the name of a firewall after you create it.

$sel:httpStatus:AssociateFirewallPolicyResponse', associateFirewallPolicyResponse_httpStatus - The response's http status code.

Response Lenses

associateFirewallPolicyResponse_updateToken :: Lens' AssociateFirewallPolicyResponse (Maybe Text)

An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

associateFirewallPolicyResponse_firewallArn :: Lens' AssociateFirewallPolicyResponse (Maybe Text)

The Amazon Resource Name (ARN) of the firewall.

associateFirewallPolicyResponse_firewallPolicyArn :: Lens' AssociateFirewallPolicyResponse (Maybe Text)

The Amazon Resource Name (ARN) of the firewall policy.

associateFirewallPolicyResponse_firewallName :: Lens' AssociateFirewallPolicyResponse (Maybe Text)

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

associateFirewallPolicyResponse_httpStatus :: Lens' AssociateFirewallPolicyResponse Int

The response's http status code.