Network.AWS.NetworkFirewall.CreateFirewall

Description

Creates an AWS Network Firewall Firewall and accompanying FirewallStatus for a VPC.

The firewall defines the configuration settings for an AWS Network Firewall firewall. The settings that you can define at creation include the firewall policy, the subnets in your VPC to use for the firewall endpoints, and any tags that are attached to the firewall AWS resource.

After you create a firewall, you can provide additional settings, like the logging configuration.

To update the settings for a firewall, you use the operations that apply to the settings themselves, for example UpdateLoggingConfiguration, AssociateSubnets, and UpdateFirewallDeleteProtection.

To manage a firewall's tags, use the standard AWS resource tagging operations, ListTagsForResource, TagResource, and UntagResource.

To retrieve information about firewalls, use ListFirewalls and DescribeFirewall.

Creating a Request

data CreateFirewall

See: newCreateFirewall smart constructor.

Constructors

CreateFirewall'

Fields

firewallPolicyChangeProtection :: Maybe Bool

A setting indicating whether the firewall is protected against a change to the firewall policy association. Use this setting to protect against accidentally modifying the firewall policy for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

subnetChangeProtection :: Maybe Bool

A setting indicating whether the firewall is protected against changes to the subnet associations. Use this setting to protect against accidentally modifying the subnet associations for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

deleteProtection :: Maybe Bool

A flag indicating whether it is possible to delete the firewall. A setting of TRUE indicates that the firewall is protected against deletion. Use this setting to protect against accidentally deleting a firewall that is in use. When you create a firewall, the operation initializes this flag to TRUE.

description :: Maybe Text

A description of the firewall.

tags :: Maybe (NonEmpty Tag)

The key:value pairs to associate with the resource.

firewallName :: Text

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

firewallPolicyArn :: Text

The Amazon Resource Name (ARN) of the FirewallPolicy that you want to use for the firewall.

vpcId :: Text

The unique identifier of the VPC where Network Firewall should create the firewall.

You can't change this setting after you create the firewall.

subnetMappings :: [SubnetMapping]

The public subnets to use for your Network Firewall firewalls. Each subnet must belong to a different Availability Zone in the VPC. Network Firewall creates a firewall endpoint in each subnet.

newCreateFirewall

Arguments

:: Text
$sel:firewallName:CreateFirewall'

-> Text
$sel:firewallPolicyArn:CreateFirewall'

-> Text
$sel:vpcId:CreateFirewall'

-> CreateFirewall

Create a value of CreateFirewall with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:firewallPolicyChangeProtection:CreateFirewall', createFirewall_firewallPolicyChangeProtection - A setting indicating whether the firewall is protected against a change to the firewall policy association. Use this setting to protect against accidentally modifying the firewall policy for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

$sel:subnetChangeProtection:CreateFirewall', createFirewall_subnetChangeProtection - A setting indicating whether the firewall is protected against changes to the subnet associations. Use this setting to protect against accidentally modifying the subnet associations for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

$sel:deleteProtection:CreateFirewall', createFirewall_deleteProtection - A flag indicating whether it is possible to delete the firewall. A setting of TRUE indicates that the firewall is protected against deletion. Use this setting to protect against accidentally deleting a firewall that is in use. When you create a firewall, the operation initializes this flag to TRUE.

$sel:description:CreateFirewall', createFirewall_description - A description of the firewall.

$sel:tags:CreateFirewall', createFirewall_tags - The key:value pairs to associate with the resource.

$sel:firewallName:CreateFirewall', createFirewall_firewallName - The descriptive name of the firewall. You can't change the name of a firewall after you create it.

$sel:firewallPolicyArn:CreateFirewall', createFirewall_firewallPolicyArn - The Amazon Resource Name (ARN) of the FirewallPolicy that you want to use for the firewall.

$sel:vpcId:CreateFirewall', createFirewall_vpcId - The unique identifier of the VPC where Network Firewall should create the firewall.

You can't change this setting after you create the firewall.

$sel:subnetMappings:CreateFirewall', createFirewall_subnetMappings - The public subnets to use for your Network Firewall firewalls. Each subnet must belong to a different Availability Zone in the VPC. Network Firewall creates a firewall endpoint in each subnet.

Request Lenses

createFirewall_firewallPolicyChangeProtection :: Lens' CreateFirewall (Maybe Bool)

A setting indicating whether the firewall is protected against a change to the firewall policy association. Use this setting to protect against accidentally modifying the firewall policy for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

createFirewall_subnetChangeProtection :: Lens' CreateFirewall (Maybe Bool)

A setting indicating whether the firewall is protected against changes to the subnet associations. Use this setting to protect against accidentally modifying the subnet associations for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

createFirewall_deleteProtection :: Lens' CreateFirewall (Maybe Bool)

A flag indicating whether it is possible to delete the firewall. A setting of TRUE indicates that the firewall is protected against deletion. Use this setting to protect against accidentally deleting a firewall that is in use. When you create a firewall, the operation initializes this flag to TRUE.

createFirewall_description :: Lens' CreateFirewall (Maybe Text)

A description of the firewall.

createFirewall_tags :: Lens' CreateFirewall (Maybe (NonEmpty Tag))

The key:value pairs to associate with the resource.

createFirewall_firewallName :: Lens' CreateFirewall Text

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

createFirewall_firewallPolicyArn :: Lens' CreateFirewall Text

The Amazon Resource Name (ARN) of the FirewallPolicy that you want to use for the firewall.

createFirewall_vpcId :: Lens' CreateFirewall Text

The unique identifier of the VPC where Network Firewall should create the firewall.

You can't change this setting after you create the firewall.

createFirewall_subnetMappings :: Lens' CreateFirewall [SubnetMapping]

The public subnets to use for your Network Firewall firewalls. Each subnet must belong to a different Availability Zone in the VPC. Network Firewall creates a firewall endpoint in each subnet.

Destructuring the Response

data CreateFirewallResponse

See: newCreateFirewallResponse smart constructor.

Constructors

CreateFirewallResponse'

Fields

firewallStatus :: Maybe FirewallStatus

Detailed information about the current status of a Firewall. You can retrieve this for a firewall by calling DescribeFirewall and providing the firewall name and ARN.

firewall :: Maybe Firewall

The configuration settings for the firewall. These settings include the firewall policy and the subnets in your VPC to use for the firewall endpoints.

httpStatus :: Int

The response's http status code.

newCreateFirewallResponse

Arguments

:: Int
$sel:httpStatus:CreateFirewallResponse'

-> CreateFirewallResponse

Create a value of CreateFirewallResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:firewallStatus:CreateFirewallResponse', createFirewallResponse_firewallStatus - Detailed information about the current status of a Firewall. You can retrieve this for a firewall by calling DescribeFirewall and providing the firewall name and ARN.

$sel:firewall:CreateFirewallResponse', createFirewallResponse_firewall - The configuration settings for the firewall. These settings include the firewall policy and the subnets in your VPC to use for the firewall endpoints.

$sel:httpStatus:CreateFirewallResponse', createFirewallResponse_httpStatus - The response's http status code.

Response Lenses

createFirewallResponse_firewallStatus :: Lens' CreateFirewallResponse (Maybe FirewallStatus)

Detailed information about the current status of a Firewall. You can retrieve this for a firewall by calling DescribeFirewall and providing the firewall name and ARN.

createFirewallResponse_firewall :: Lens' CreateFirewallResponse (Maybe Firewall)

The configuration settings for the firewall. These settings include the firewall policy and the subnets in your VPC to use for the firewall endpoints.

createFirewallResponse_httpStatus :: Lens' CreateFirewallResponse Int

The response's http status code.