Network.AWS.NetworkFirewall.CreateRuleGroup

Description

Creates the specified stateless or stateful rule group, which includes the rules for network traffic inspection, a capacity setting, and tags.

You provide your rule group specification in your request using either RuleGroup or Rules.

Creating a Request

data CreateRuleGroup

See: newCreateRuleGroup smart constructor.

Constructors

CreateRuleGroup'

Fields

rules :: Maybe Text

A string containing stateful rule group rules specifications in Suricata flat format, with one rule per line. Use this to import your existing Suricata compatible rule groups.

You must provide either this rules setting or a populated RuleGroup setting, but not both.

You can provide your rule group specification in Suricata flat format through this setting when you create or update your rule group. The call response returns a RuleGroup object that Network Firewall has populated from your string.

description :: Maybe Text

A description of the rule group.

dryRun :: Maybe Bool

Indicates whether you want Network Firewall to just check the validity of the request, rather than run the request.

If set to TRUE, Network Firewall checks whether the request can run successfully, but doesn't actually make the requested changes. The call returns the value that the request would return if you ran it with dry run set to FALSE, but doesn't make additions or changes to your resources. This option allows you to make sure that you have the required permissions to run the request and that your request parameters are valid.

If set to FALSE, Network Firewall makes the requested changes to your resources.

tags :: Maybe (NonEmpty Tag)

The key:value pairs to associate with the resource.

ruleGroup :: Maybe RuleGroup

An object that defines the rule group rules.

You must provide either this rule group setting or a Rules setting, but not both.

ruleGroupName :: Text

The descriptive name of the rule group. You can't change the name of a rule group after you create it.

type' :: RuleGroupType

Indicates whether the rule group is stateless or stateful. If the rule group is stateless, it contains stateless rules. If it is stateful, it contains stateful rules.

capacity :: Int

The maximum operating resources that this rule group can use. Rule group capacity is fixed at creation. When you update a rule group, you are limited to this capacity. When you reference a rule group from a firewall policy, Network Firewall reserves this capacity for the rule group.

You can retrieve the capacity that would be required for a rule group before you create the rule group by calling CreateRuleGroup with DryRun set to TRUE.

You can't change or exceed this capacity when you update the rule group, so leave room for your rule group to grow.

Capacity for a stateless rule group

For a stateless rule group, the capacity required is the sum of the capacity requirements of the individual rules that you expect to have in the rule group.

To calculate the capacity requirement of a single rule, multiply the capacity requirement values of each of the rule's match settings:

A match setting with no criteria specified has a value of 1.

A match setting with Any specified has a value of 1.

All other match settings have a value equal to the number of elements provided in the setting. For example, a protocol setting ["UDP"] and a source setting ["10.0.0.0/24"] each have a value of 1. A protocol setting ["UDP","TCP"] has a value of 2. A source setting ["10.0.0.0/24","10.0.0.1/24","10.0.0.2/24"] has a value of 3.

A rule with no criteria specified in any of its match settings has a capacity requirement of 1. A rule with protocol setting ["UDP","TCP"], source setting ["10.0.0.0/24","10.0.0.1/24","10.0.0.2/24"], and a single specification or no specification for each of the other match settings has a capacity requirement of 6.

Capacity for a stateful rule group

For a stateful rule group, the minimum capacity required is the number of individual rules that you expect to have in the rule group.

newCreateRuleGroup

Arguments

:: Text
$sel:ruleGroupName:CreateRuleGroup'

-> RuleGroupType
$sel:type':CreateRuleGroup'

-> Int
$sel:capacity:CreateRuleGroup'

-> CreateRuleGroup

Create a value of CreateRuleGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rules:CreateRuleGroup', createRuleGroup_rules - A string containing stateful rule group rules specifications in Suricata flat format, with one rule per line. Use this to import your existing Suricata compatible rule groups.

You must provide either this rules setting or a populated RuleGroup setting, but not both.

You can provide your rule group specification in Suricata flat format through this setting when you create or update your rule group. The call response returns a RuleGroup object that Network Firewall has populated from your string.

$sel:description:CreateRuleGroup', createRuleGroup_description - A description of the rule group.

$sel:dryRun:CreateRuleGroup', createRuleGroup_dryRun - Indicates whether you want Network Firewall to just check the validity of the request, rather than run the request.

If set to TRUE, Network Firewall checks whether the request can run successfully, but doesn't actually make the requested changes. The call returns the value that the request would return if you ran it with dry run set to FALSE, but doesn't make additions or changes to your resources. This option allows you to make sure that you have the required permissions to run the request and that your request parameters are valid.

If set to FALSE, Network Firewall makes the requested changes to your resources.

$sel:tags:CreateRuleGroup', createRuleGroup_tags - The key:value pairs to associate with the resource.

$sel:ruleGroup:CreateRuleGroup', createRuleGroup_ruleGroup - An object that defines the rule group rules.

You must provide either this rule group setting or a Rules setting, but not both.

$sel:ruleGroupName:CreateRuleGroup', createRuleGroup_ruleGroupName - The descriptive name of the rule group. You can't change the name of a rule group after you create it.

$sel:type':CreateRuleGroup', createRuleGroup_type - Indicates whether the rule group is stateless or stateful. If the rule group is stateless, it contains stateless rules. If it is stateful, it contains stateful rules.

$sel:capacity:CreateRuleGroup', createRuleGroup_capacity - The maximum operating resources that this rule group can use. Rule group capacity is fixed at creation. When you update a rule group, you are limited to this capacity. When you reference a rule group from a firewall policy, Network Firewall reserves this capacity for the rule group.

You can retrieve the capacity that would be required for a rule group before you create the rule group by calling CreateRuleGroup with DryRun set to TRUE.

You can't change or exceed this capacity when you update the rule group, so leave room for your rule group to grow.

Capacity for a stateless rule group

For a stateless rule group, the capacity required is the sum of the capacity requirements of the individual rules that you expect to have in the rule group.

To calculate the capacity requirement of a single rule, multiply the capacity requirement values of each of the rule's match settings:

A match setting with no criteria specified has a value of 1.

A match setting with Any specified has a value of 1.

All other match settings have a value equal to the number of elements provided in the setting. For example, a protocol setting ["UDP"] and a source setting ["10.0.0.0/24"] each have a value of 1. A protocol setting ["UDP","TCP"] has a value of 2. A source setting ["10.0.0.0/24","10.0.0.1/24","10.0.0.2/24"] has a value of 3.

A rule with no criteria specified in any of its match settings has a capacity requirement of 1. A rule with protocol setting ["UDP","TCP"], source setting ["10.0.0.0/24","10.0.0.1/24","10.0.0.2/24"], and a single specification or no specification for each of the other match settings has a capacity requirement of 6.

Capacity for a stateful rule group

For a stateful rule group, the minimum capacity required is the number of individual rules that you expect to have in the rule group.

Request Lenses

createRuleGroup_rules :: Lens' CreateRuleGroup (Maybe Text)

A string containing stateful rule group rules specifications in Suricata flat format, with one rule per line. Use this to import your existing Suricata compatible rule groups.

You must provide either this rules setting or a populated RuleGroup setting, but not both.

You can provide your rule group specification in Suricata flat format through this setting when you create or update your rule group. The call response returns a RuleGroup object that Network Firewall has populated from your string.

createRuleGroup_description :: Lens' CreateRuleGroup (Maybe Text)

A description of the rule group.

createRuleGroup_dryRun :: Lens' CreateRuleGroup (Maybe Bool)

Indicates whether you want Network Firewall to just check the validity of the request, rather than run the request.

If set to TRUE, Network Firewall checks whether the request can run successfully, but doesn't actually make the requested changes. The call returns the value that the request would return if you ran it with dry run set to FALSE, but doesn't make additions or changes to your resources. This option allows you to make sure that you have the required permissions to run the request and that your request parameters are valid.

If set to FALSE, Network Firewall makes the requested changes to your resources.

createRuleGroup_tags :: Lens' CreateRuleGroup (Maybe (NonEmpty Tag))

The key:value pairs to associate with the resource.

createRuleGroup_ruleGroup :: Lens' CreateRuleGroup (Maybe RuleGroup)

An object that defines the rule group rules.

You must provide either this rule group setting or a Rules setting, but not both.

createRuleGroup_ruleGroupName :: Lens' CreateRuleGroup Text

The descriptive name of the rule group. You can't change the name of a rule group after you create it.

createRuleGroup_type :: Lens' CreateRuleGroup RuleGroupType

Indicates whether the rule group is stateless or stateful. If the rule group is stateless, it contains stateless rules. If it is stateful, it contains stateful rules.

createRuleGroup_capacity :: Lens' CreateRuleGroup Int

The maximum operating resources that this rule group can use. Rule group capacity is fixed at creation. When you update a rule group, you are limited to this capacity. When you reference a rule group from a firewall policy, Network Firewall reserves this capacity for the rule group.

You can retrieve the capacity that would be required for a rule group before you create the rule group by calling CreateRuleGroup with DryRun set to TRUE.

You can't change or exceed this capacity when you update the rule group, so leave room for your rule group to grow.

Capacity for a stateless rule group

For a stateless rule group, the capacity required is the sum of the capacity requirements of the individual rules that you expect to have in the rule group.

To calculate the capacity requirement of a single rule, multiply the capacity requirement values of each of the rule's match settings:

A match setting with no criteria specified has a value of 1.

A match setting with Any specified has a value of 1.

All other match settings have a value equal to the number of elements provided in the setting. For example, a protocol setting ["UDP"] and a source setting ["10.0.0.0/24"] each have a value of 1. A protocol setting ["UDP","TCP"] has a value of 2. A source setting ["10.0.0.0/24","10.0.0.1/24","10.0.0.2/24"] has a value of 3.

A rule with no criteria specified in any of its match settings has a capacity requirement of 1. A rule with protocol setting ["UDP","TCP"], source setting ["10.0.0.0/24","10.0.0.1/24","10.0.0.2/24"], and a single specification or no specification for each of the other match settings has a capacity requirement of 6.

Capacity for a stateful rule group

For a stateful rule group, the minimum capacity required is the number of individual rules that you expect to have in the rule group.

Destructuring the Response

data CreateRuleGroupResponse

See: newCreateRuleGroupResponse smart constructor.

Constructors

CreateRuleGroupResponse'

Fields

httpStatus :: Int

The response's http status code.

updateToken :: Text

A token used for optimistic locking. Network Firewall returns a token to your requests that access the rule group. The token marks the state of the rule group resource at the time of the request.

To make changes to the rule group, you provide the token in your request. Network Firewall uses the token to ensure that the rule group hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the rule group again to get a current copy of it with a current token. Reapply your changes as needed, then try the operation again using the new token.

ruleGroupResponse :: RuleGroupResponse

The high-level properties of a rule group. This, along with the RuleGroup, define the rule group. You can retrieve all objects for a rule group by calling DescribeRuleGroup.

newCreateRuleGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateRuleGroupResponse'

-> Text
$sel:updateToken:CreateRuleGroupResponse'

-> RuleGroupResponse
$sel:ruleGroupResponse:CreateRuleGroupResponse'

-> CreateRuleGroupResponse

Create a value of CreateRuleGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateRuleGroupResponse', createRuleGroupResponse_httpStatus - The response's http status code.

$sel:updateToken:CreateRuleGroupResponse', createRuleGroupResponse_updateToken - A token used for optimistic locking. Network Firewall returns a token to your requests that access the rule group. The token marks the state of the rule group resource at the time of the request.

To make changes to the rule group, you provide the token in your request. Network Firewall uses the token to ensure that the rule group hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the rule group again to get a current copy of it with a current token. Reapply your changes as needed, then try the operation again using the new token.

$sel:ruleGroupResponse:CreateRuleGroupResponse', createRuleGroupResponse_ruleGroupResponse - The high-level properties of a rule group. This, along with the RuleGroup, define the rule group. You can retrieve all objects for a rule group by calling DescribeRuleGroup.

Response Lenses

createRuleGroupResponse_httpStatus :: Lens' CreateRuleGroupResponse Int

The response's http status code.

createRuleGroupResponse_updateToken :: Lens' CreateRuleGroupResponse Text

A token used for optimistic locking. Network Firewall returns a token to your requests that access the rule group. The token marks the state of the rule group resource at the time of the request.

To make changes to the rule group, you provide the token in your request. Network Firewall uses the token to ensure that the rule group hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the rule group again to get a current copy of it with a current token. Reapply your changes as needed, then try the operation again using the new token.

createRuleGroupResponse_ruleGroupResponse :: Lens' CreateRuleGroupResponse RuleGroupResponse

The high-level properties of a rule group. This, along with the RuleGroup, define the rule group. You can retrieve all objects for a rule group by calling DescribeRuleGroup.