Network.AWS.NetworkFirewall.DeleteFirewall

Description

Deletes the specified Firewall and its FirewallStatus. This operation requires the firewall's DeleteProtection flag to be FALSE. You can't revert this operation.

You can check whether a firewall is in use by reviewing the route tables for the Availability Zones where you have firewall subnet mappings. Retrieve the subnet mappings by calling DescribeFirewall. You define and update the route tables through Amazon VPC. As needed, update the route tables for the zones to remove the firewall endpoints. When the route tables no longer use the firewall endpoints, you can remove the firewall safely.

To delete a firewall, remove the delete protection if you need to using UpdateFirewallDeleteProtection, then delete the firewall by calling DeleteFirewall.

Creating a Request

data DeleteFirewall

See: newDeleteFirewall smart constructor.

Constructors

DeleteFirewall'

Fields

firewallArn :: Maybe Text

The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.

firewallName :: Maybe Text

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

You must specify the ARN or the name, and you can specify both.

newDeleteFirewall :: DeleteFirewall

Create a value of DeleteFirewall with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:firewallArn:DeleteFirewall', deleteFirewall_firewallArn - The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.

$sel:firewallName:DeleteFirewall', deleteFirewall_firewallName - The descriptive name of the firewall. You can't change the name of a firewall after you create it.

You must specify the ARN or the name, and you can specify both.

Request Lenses

deleteFirewall_firewallArn :: Lens' DeleteFirewall (Maybe Text)

The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.

deleteFirewall_firewallName :: Lens' DeleteFirewall (Maybe Text)

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

You must specify the ARN or the name, and you can specify both.

Destructuring the Response

data DeleteFirewallResponse

See: newDeleteFirewallResponse smart constructor.

Constructors

DeleteFirewallResponse'

Fields

firewallStatus :: Maybe FirewallStatus

firewall :: Maybe Firewall

httpStatus :: Int

The response's http status code.

newDeleteFirewallResponse

Arguments

:: Int
$sel:httpStatus:DeleteFirewallResponse'

-> DeleteFirewallResponse

Create a value of DeleteFirewallResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:firewallStatus:DeleteFirewallResponse', deleteFirewallResponse_firewallStatus - Undocumented member.

$sel:firewall:DeleteFirewallResponse', deleteFirewallResponse_firewall - Undocumented member.

$sel:httpStatus:DeleteFirewallResponse', deleteFirewallResponse_httpStatus - The response's http status code.

Response Lenses

deleteFirewallResponse_firewallStatus :: Lens' DeleteFirewallResponse (Maybe FirewallStatus)

Undocumented member.

deleteFirewallResponse_firewall :: Lens' DeleteFirewallResponse (Maybe Firewall)

Undocumented member.

deleteFirewallResponse_httpStatus :: Lens' DeleteFirewallResponse Int

The response's http status code.