Network.AWS.NetworkFirewall.DeleteFirewallPolicy

Description

Deletes the specified FirewallPolicy.

Creating a Request

data DeleteFirewallPolicy

See: newDeleteFirewallPolicy smart constructor.

Constructors

DeleteFirewallPolicy'

Fields

firewallPolicyName :: Maybe Text

The descriptive name of the firewall policy. You can't change the name of a firewall policy after you create it.

You must specify the ARN or the name, and you can specify both.

firewallPolicyArn :: Maybe Text

The Amazon Resource Name (ARN) of the firewall policy.

You must specify the ARN or the name, and you can specify both.

newDeleteFirewallPolicy :: DeleteFirewallPolicy

Create a value of DeleteFirewallPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:firewallPolicyName:DeleteFirewallPolicy', deleteFirewallPolicy_firewallPolicyName - The descriptive name of the firewall policy. You can't change the name of a firewall policy after you create it.

You must specify the ARN or the name, and you can specify both.

$sel:firewallPolicyArn:DeleteFirewallPolicy', deleteFirewallPolicy_firewallPolicyArn - The Amazon Resource Name (ARN) of the firewall policy.

You must specify the ARN or the name, and you can specify both.

Request Lenses

deleteFirewallPolicy_firewallPolicyName :: Lens' DeleteFirewallPolicy (Maybe Text)

The descriptive name of the firewall policy. You can't change the name of a firewall policy after you create it.

You must specify the ARN or the name, and you can specify both.

deleteFirewallPolicy_firewallPolicyArn :: Lens' DeleteFirewallPolicy (Maybe Text)

The Amazon Resource Name (ARN) of the firewall policy.

You must specify the ARN or the name, and you can specify both.

Destructuring the Response

data DeleteFirewallPolicyResponse

See: newDeleteFirewallPolicyResponse smart constructor.

Constructors

DeleteFirewallPolicyResponse'

Fields

httpStatus :: Int

The response's http status code.

firewallPolicyResponse :: FirewallPolicyResponse

The object containing the definition of the FirewallPolicyResponse that you asked to delete.

newDeleteFirewallPolicyResponse

Arguments

:: Int
$sel:httpStatus:DeleteFirewallPolicyResponse'

-> FirewallPolicyResponse
$sel:firewallPolicyResponse:DeleteFirewallPolicyResponse'

-> DeleteFirewallPolicyResponse

Create a value of DeleteFirewallPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteFirewallPolicyResponse', deleteFirewallPolicyResponse_httpStatus - The response's http status code.

$sel:firewallPolicyResponse:DeleteFirewallPolicyResponse', deleteFirewallPolicyResponse_firewallPolicyResponse - The object containing the definition of the FirewallPolicyResponse that you asked to delete.

Response Lenses

deleteFirewallPolicyResponse_httpStatus :: Lens' DeleteFirewallPolicyResponse Int

The response's http status code.

deleteFirewallPolicyResponse_firewallPolicyResponse :: Lens' DeleteFirewallPolicyResponse FirewallPolicyResponse

The object containing the definition of the FirewallPolicyResponse that you asked to delete.