Network.AWS.NetworkFirewall.DeleteRuleGroup

Description

Deletes the specified RuleGroup.

Creating a Request

data DeleteRuleGroup

See: newDeleteRuleGroup smart constructor.

Constructors

DeleteRuleGroup'

Fields

ruleGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of the rule group.

You must specify the ARN or the name, and you can specify both.

type' :: Maybe RuleGroupType

Indicates whether the rule group is stateless or stateful. If the rule group is stateless, it contains stateless rules. If it is stateful, it contains stateful rules.

This setting is required for requests that do not include the RuleGroupARN.

ruleGroupName :: Maybe Text

The descriptive name of the rule group. You can't change the name of a rule group after you create it.

You must specify the ARN or the name, and you can specify both.

newDeleteRuleGroup :: DeleteRuleGroup

Create a value of DeleteRuleGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleGroupArn:DeleteRuleGroup', deleteRuleGroup_ruleGroupArn - The Amazon Resource Name (ARN) of the rule group.

You must specify the ARN or the name, and you can specify both.

$sel:type':DeleteRuleGroup', deleteRuleGroup_type - Indicates whether the rule group is stateless or stateful. If the rule group is stateless, it contains stateless rules. If it is stateful, it contains stateful rules.

This setting is required for requests that do not include the RuleGroupARN.

$sel:ruleGroupName:DeleteRuleGroup', deleteRuleGroup_ruleGroupName - The descriptive name of the rule group. You can't change the name of a rule group after you create it.

You must specify the ARN or the name, and you can specify both.

Request Lenses

deleteRuleGroup_ruleGroupArn :: Lens' DeleteRuleGroup (Maybe Text)

The Amazon Resource Name (ARN) of the rule group.

You must specify the ARN or the name, and you can specify both.

deleteRuleGroup_type :: Lens' DeleteRuleGroup (Maybe RuleGroupType)

Indicates whether the rule group is stateless or stateful. If the rule group is stateless, it contains stateless rules. If it is stateful, it contains stateful rules.

This setting is required for requests that do not include the RuleGroupARN.

deleteRuleGroup_ruleGroupName :: Lens' DeleteRuleGroup (Maybe Text)

The descriptive name of the rule group. You can't change the name of a rule group after you create it.

You must specify the ARN or the name, and you can specify both.

Destructuring the Response

data DeleteRuleGroupResponse

See: newDeleteRuleGroupResponse smart constructor.

Constructors

DeleteRuleGroupResponse'

Fields

httpStatus :: Int

The response's http status code.

ruleGroupResponse :: RuleGroupResponse

The high-level properties of a rule group. This, along with the RuleGroup, define the rule group. You can retrieve all objects for a rule group by calling DescribeRuleGroup.

newDeleteRuleGroupResponse

Arguments

:: Int
$sel:httpStatus:DeleteRuleGroupResponse'

-> RuleGroupResponse
$sel:ruleGroupResponse:DeleteRuleGroupResponse'

-> DeleteRuleGroupResponse

Create a value of DeleteRuleGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteRuleGroupResponse', deleteRuleGroupResponse_httpStatus - The response's http status code.

$sel:ruleGroupResponse:DeleteRuleGroupResponse', deleteRuleGroupResponse_ruleGroupResponse - The high-level properties of a rule group. This, along with the RuleGroup, define the rule group. You can retrieve all objects for a rule group by calling DescribeRuleGroup.

Response Lenses

deleteRuleGroupResponse_httpStatus :: Lens' DeleteRuleGroupResponse Int

The response's http status code.

deleteRuleGroupResponse_ruleGroupResponse :: Lens' DeleteRuleGroupResponse RuleGroupResponse

The high-level properties of a rule group. This, along with the RuleGroup, define the rule group. You can retrieve all objects for a rule group by calling DescribeRuleGroup.