Network.AWS.NetworkFirewall.DescribeFirewallPolicy

Description

Returns the data objects for the specified firewall policy.

Creating a Request

data DescribeFirewallPolicy

See: newDescribeFirewallPolicy smart constructor.

Constructors

DescribeFirewallPolicy'

Fields

firewallPolicyName :: Maybe Text

The descriptive name of the firewall policy. You can't change the name of a firewall policy after you create it.

You must specify the ARN or the name, and you can specify both.

firewallPolicyArn :: Maybe Text

The Amazon Resource Name (ARN) of the firewall policy.

You must specify the ARN or the name, and you can specify both.

newDescribeFirewallPolicy :: DescribeFirewallPolicy

Create a value of DescribeFirewallPolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:firewallPolicyName:DescribeFirewallPolicy', describeFirewallPolicy_firewallPolicyName - The descriptive name of the firewall policy. You can't change the name of a firewall policy after you create it.

You must specify the ARN or the name, and you can specify both.

$sel:firewallPolicyArn:DescribeFirewallPolicy', describeFirewallPolicy_firewallPolicyArn - The Amazon Resource Name (ARN) of the firewall policy.

You must specify the ARN or the name, and you can specify both.

Request Lenses

describeFirewallPolicy_firewallPolicyName :: Lens' DescribeFirewallPolicy (Maybe Text)

The descriptive name of the firewall policy. You can't change the name of a firewall policy after you create it.

You must specify the ARN or the name, and you can specify both.

describeFirewallPolicy_firewallPolicyArn :: Lens' DescribeFirewallPolicy (Maybe Text)

The Amazon Resource Name (ARN) of the firewall policy.

You must specify the ARN or the name, and you can specify both.

Destructuring the Response

data DescribeFirewallPolicyResponse

See: newDescribeFirewallPolicyResponse smart constructor.

Constructors

DescribeFirewallPolicyResponse'

Fields

firewallPolicy :: Maybe FirewallPolicy

The policy for the specified firewall policy.

httpStatus :: Int

The response's http status code.

updateToken :: Text

A token used for optimistic locking. Network Firewall returns a token to your requests that access the firewall policy. The token marks the state of the policy resource at the time of the request.

To make changes to the policy, you provide the token in your request. Network Firewall uses the token to ensure that the policy hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall policy again to get a current copy of it with current token. Reapply your changes as needed, then try the operation again using the new token.

firewallPolicyResponse :: FirewallPolicyResponse

The high-level properties of a firewall policy. This, along with the FirewallPolicy, define the policy. You can retrieve all objects for a firewall policy by calling DescribeFirewallPolicy.

newDescribeFirewallPolicyResponse

Arguments

:: Int
$sel:httpStatus:DescribeFirewallPolicyResponse'

-> Text
$sel:updateToken:DescribeFirewallPolicyResponse'

-> FirewallPolicyResponse
$sel:firewallPolicyResponse:DescribeFirewallPolicyResponse'

-> DescribeFirewallPolicyResponse

Create a value of DescribeFirewallPolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:firewallPolicy:DescribeFirewallPolicyResponse', describeFirewallPolicyResponse_firewallPolicy - The policy for the specified firewall policy.

$sel:httpStatus:DescribeFirewallPolicyResponse', describeFirewallPolicyResponse_httpStatus - The response's http status code.

$sel:updateToken:DescribeFirewallPolicyResponse', describeFirewallPolicyResponse_updateToken - A token used for optimistic locking. Network Firewall returns a token to your requests that access the firewall policy. The token marks the state of the policy resource at the time of the request.

To make changes to the policy, you provide the token in your request. Network Firewall uses the token to ensure that the policy hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall policy again to get a current copy of it with current token. Reapply your changes as needed, then try the operation again using the new token.

$sel:firewallPolicyResponse:DescribeFirewallPolicyResponse', describeFirewallPolicyResponse_firewallPolicyResponse - The high-level properties of a firewall policy. This, along with the FirewallPolicy, define the policy. You can retrieve all objects for a firewall policy by calling DescribeFirewallPolicy.

Response Lenses

describeFirewallPolicyResponse_firewallPolicy :: Lens' DescribeFirewallPolicyResponse (Maybe FirewallPolicy)

The policy for the specified firewall policy.

describeFirewallPolicyResponse_httpStatus :: Lens' DescribeFirewallPolicyResponse Int

The response's http status code.

describeFirewallPolicyResponse_updateToken :: Lens' DescribeFirewallPolicyResponse Text

A token used for optimistic locking. Network Firewall returns a token to your requests that access the firewall policy. The token marks the state of the policy resource at the time of the request.

To make changes to the policy, you provide the token in your request. Network Firewall uses the token to ensure that the policy hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall policy again to get a current copy of it with current token. Reapply your changes as needed, then try the operation again using the new token.

describeFirewallPolicyResponse_firewallPolicyResponse :: Lens' DescribeFirewallPolicyResponse FirewallPolicyResponse

The high-level properties of a firewall policy. This, along with the FirewallPolicy, define the policy. You can retrieve all objects for a firewall policy by calling DescribeFirewallPolicy.