Network.AWS.NetworkFirewall.DisassociateSubnets

Description

Removes the specified subnet associations from the firewall. This removes the firewall endpoints from the subnets and removes any network filtering protections that the endpoints were providing.

Creating a Request

data DisassociateSubnets

See: newDisassociateSubnets smart constructor.

Constructors

DisassociateSubnets'

Fields

updateToken :: Maybe Text

An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

firewallArn :: Maybe Text

The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.

firewallName :: Maybe Text

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

You must specify the ARN or the name, and you can specify both.

subnetIds :: [Text]

The unique identifiers for the subnets that you want to disassociate.

newDisassociateSubnets :: DisassociateSubnets

Create a value of DisassociateSubnets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateToken:DisassociateSubnets', disassociateSubnets_updateToken - An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

$sel:firewallArn:DisassociateSubnets', disassociateSubnets_firewallArn - The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.

$sel:firewallName:DisassociateSubnets', disassociateSubnets_firewallName - The descriptive name of the firewall. You can't change the name of a firewall after you create it.

You must specify the ARN or the name, and you can specify both.

$sel:subnetIds:DisassociateSubnets', disassociateSubnets_subnetIds - The unique identifiers for the subnets that you want to disassociate.

Request Lenses

disassociateSubnets_updateToken :: Lens' DisassociateSubnets (Maybe Text)

An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

disassociateSubnets_firewallArn :: Lens' DisassociateSubnets (Maybe Text)

The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.

disassociateSubnets_firewallName :: Lens' DisassociateSubnets (Maybe Text)

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

You must specify the ARN or the name, and you can specify both.

disassociateSubnets_subnetIds :: Lens' DisassociateSubnets [Text]

The unique identifiers for the subnets that you want to disassociate.

Destructuring the Response

data DisassociateSubnetsResponse

See: newDisassociateSubnetsResponse smart constructor.

Constructors

DisassociateSubnetsResponse'

Fields

subnetMappings :: Maybe [SubnetMapping]

The IDs of the subnets that are associated with the firewall.

updateToken :: Maybe Text

An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

firewallArn :: Maybe Text

The Amazon Resource Name (ARN) of the firewall.

firewallName :: Maybe Text

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

httpStatus :: Int

The response's http status code.

newDisassociateSubnetsResponse

Arguments

:: Int
$sel:httpStatus:DisassociateSubnetsResponse'

-> DisassociateSubnetsResponse

Create a value of DisassociateSubnetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subnetMappings:DisassociateSubnetsResponse', disassociateSubnetsResponse_subnetMappings - The IDs of the subnets that are associated with the firewall.

$sel:updateToken:DisassociateSubnets', disassociateSubnetsResponse_updateToken - An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

$sel:firewallArn:DisassociateSubnets', disassociateSubnetsResponse_firewallArn - The Amazon Resource Name (ARN) of the firewall.

$sel:firewallName:DisassociateSubnets', disassociateSubnetsResponse_firewallName - The descriptive name of the firewall. You can't change the name of a firewall after you create it.

$sel:httpStatus:DisassociateSubnetsResponse', disassociateSubnetsResponse_httpStatus - The response's http status code.

Response Lenses

disassociateSubnetsResponse_subnetMappings :: Lens' DisassociateSubnetsResponse (Maybe [SubnetMapping])

The IDs of the subnets that are associated with the firewall.

disassociateSubnetsResponse_updateToken :: Lens' DisassociateSubnetsResponse (Maybe Text)

An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

disassociateSubnetsResponse_firewallArn :: Lens' DisassociateSubnetsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the firewall.

disassociateSubnetsResponse_firewallName :: Lens' DisassociateSubnetsResponse (Maybe Text)

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

disassociateSubnetsResponse_httpStatus :: Lens' DisassociateSubnetsResponse Int

The response's http status code.