Network.AWS.NetworkFirewall.ListFirewallPolicies

Description

Retrieves the metadata for the firewall policies that you have defined. Depending on your setting for max results and the number of firewall policies, a single call might not return the full list.

This operation returns paginated results.

Creating a Request

data ListFirewallPolicies

See: newListFirewallPolicies smart constructor.

Constructors

ListFirewallPolicies'

Fields

nextToken :: Maybe Text

When you request a list of objects with a MaxResults setting, if the number of objects that are still available for retrieval exceeds the maximum you requested, Network Firewall returns a NextToken value in the response. To retrieve the next batch of objects, use the token returned from the prior request in your next request.

maxResults :: Maybe Natural

The maximum number of objects that you want Network Firewall to return for this request. If more objects are available, in the response, Network Firewall provides a NextToken value that you can use in a subsequent call to get the next batch of objects.

newListFirewallPolicies :: ListFirewallPolicies

Create a value of ListFirewallPolicies with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFirewallPolicies', listFirewallPolicies_nextToken - When you request a list of objects with a MaxResults setting, if the number of objects that are still available for retrieval exceeds the maximum you requested, Network Firewall returns a NextToken value in the response. To retrieve the next batch of objects, use the token returned from the prior request in your next request.

$sel:maxResults:ListFirewallPolicies', listFirewallPolicies_maxResults - The maximum number of objects that you want Network Firewall to return for this request. If more objects are available, in the response, Network Firewall provides a NextToken value that you can use in a subsequent call to get the next batch of objects.

Request Lenses

listFirewallPolicies_nextToken :: Lens' ListFirewallPolicies (Maybe Text)

When you request a list of objects with a MaxResults setting, if the number of objects that are still available for retrieval exceeds the maximum you requested, Network Firewall returns a NextToken value in the response. To retrieve the next batch of objects, use the token returned from the prior request in your next request.

listFirewallPolicies_maxResults :: Lens' ListFirewallPolicies (Maybe Natural)

The maximum number of objects that you want Network Firewall to return for this request. If more objects are available, in the response, Network Firewall provides a NextToken value that you can use in a subsequent call to get the next batch of objects.

Destructuring the Response

data ListFirewallPoliciesResponse

See: newListFirewallPoliciesResponse smart constructor.

Constructors

ListFirewallPoliciesResponse'

Fields

firewallPolicies :: Maybe [FirewallPolicyMetadata]

The metadata for the firewall policies. Depending on your setting for max results and the number of firewall policies that you have, this might not be the full list.

nextToken :: Maybe Text

When you request a list of objects with a MaxResults setting, if the number of objects that are still available for retrieval exceeds the maximum you requested, Network Firewall returns a NextToken value in the response. To retrieve the next batch of objects, use the token returned from the prior request in your next request.

httpStatus :: Int

The response's http status code.

newListFirewallPoliciesResponse

Arguments

:: Int
$sel:httpStatus:ListFirewallPoliciesResponse'

-> ListFirewallPoliciesResponse

Create a value of ListFirewallPoliciesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:firewallPolicies:ListFirewallPoliciesResponse', listFirewallPoliciesResponse_firewallPolicies - The metadata for the firewall policies. Depending on your setting for max results and the number of firewall policies that you have, this might not be the full list.

$sel:nextToken:ListFirewallPolicies', listFirewallPoliciesResponse_nextToken - When you request a list of objects with a MaxResults setting, if the number of objects that are still available for retrieval exceeds the maximum you requested, Network Firewall returns a NextToken value in the response. To retrieve the next batch of objects, use the token returned from the prior request in your next request.

$sel:httpStatus:ListFirewallPoliciesResponse', listFirewallPoliciesResponse_httpStatus - The response's http status code.

Response Lenses

listFirewallPoliciesResponse_firewallPolicies :: Lens' ListFirewallPoliciesResponse (Maybe [FirewallPolicyMetadata])

The metadata for the firewall policies. Depending on your setting for max results and the number of firewall policies that you have, this might not be the full list.

listFirewallPoliciesResponse_nextToken :: Lens' ListFirewallPoliciesResponse (Maybe Text)

When you request a list of objects with a MaxResults setting, if the number of objects that are still available for retrieval exceeds the maximum you requested, Network Firewall returns a NextToken value in the response. To retrieve the next batch of objects, use the token returned from the prior request in your next request.

listFirewallPoliciesResponse_httpStatus :: Lens' ListFirewallPoliciesResponse Int

The response's http status code.