Network.AWS.NetworkFirewall.ListFirewalls

Description

Retrieves the metadata for the firewalls that you have defined. If you provide VPC identifiers in your request, this returns only the firewalls for those VPCs.

Depending on your setting for max results and the number of firewalls, a single call might not return the full list.

This operation returns paginated results.

Creating a Request

data ListFirewalls

See: newListFirewalls smart constructor.

Constructors

ListFirewalls'

Fields

nextToken :: Maybe Text

When you request a list of objects with a MaxResults setting, if the number of objects that are still available for retrieval exceeds the maximum you requested, Network Firewall returns a NextToken value in the response. To retrieve the next batch of objects, use the token returned from the prior request in your next request.

vpcIds :: Maybe [Text]

The unique identifiers of the VPCs that you want Network Firewall to retrieve the firewalls for. Leave this blank to retrieve all firewalls that you have defined.

maxResults :: Maybe Natural

The maximum number of objects that you want Network Firewall to return for this request. If more objects are available, in the response, Network Firewall provides a NextToken value that you can use in a subsequent call to get the next batch of objects.

newListFirewalls :: ListFirewalls

Create a value of ListFirewalls with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFirewalls', listFirewalls_nextToken - When you request a list of objects with a MaxResults setting, if the number of objects that are still available for retrieval exceeds the maximum you requested, Network Firewall returns a NextToken value in the response. To retrieve the next batch of objects, use the token returned from the prior request in your next request.

$sel:vpcIds:ListFirewalls', listFirewalls_vpcIds - The unique identifiers of the VPCs that you want Network Firewall to retrieve the firewalls for. Leave this blank to retrieve all firewalls that you have defined.

$sel:maxResults:ListFirewalls', listFirewalls_maxResults - The maximum number of objects that you want Network Firewall to return for this request. If more objects are available, in the response, Network Firewall provides a NextToken value that you can use in a subsequent call to get the next batch of objects.

Request Lenses

listFirewalls_nextToken :: Lens' ListFirewalls (Maybe Text)

When you request a list of objects with a MaxResults setting, if the number of objects that are still available for retrieval exceeds the maximum you requested, Network Firewall returns a NextToken value in the response. To retrieve the next batch of objects, use the token returned from the prior request in your next request.

listFirewalls_vpcIds :: Lens' ListFirewalls (Maybe [Text])

The unique identifiers of the VPCs that you want Network Firewall to retrieve the firewalls for. Leave this blank to retrieve all firewalls that you have defined.

listFirewalls_maxResults :: Lens' ListFirewalls (Maybe Natural)

The maximum number of objects that you want Network Firewall to return for this request. If more objects are available, in the response, Network Firewall provides a NextToken value that you can use in a subsequent call to get the next batch of objects.

Destructuring the Response

data ListFirewallsResponse

See: newListFirewallsResponse smart constructor.

Constructors

ListFirewallsResponse'

Fields

nextToken :: Maybe Text

When you request a list of objects with a MaxResults setting, if the number of objects that are still available for retrieval exceeds the maximum you requested, Network Firewall returns a NextToken value in the response. To retrieve the next batch of objects, use the token returned from the prior request in your next request.

firewalls :: Maybe [FirewallMetadata]

The firewall metadata objects for the VPCs that you specified. Depending on your setting for max results and the number of firewalls you have, a single call might not be the full list.

httpStatus :: Int

The response's http status code.

newListFirewallsResponse

Arguments

:: Int
$sel:httpStatus:ListFirewallsResponse'

-> ListFirewallsResponse

Create a value of ListFirewallsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListFirewalls', listFirewallsResponse_nextToken - When you request a list of objects with a MaxResults setting, if the number of objects that are still available for retrieval exceeds the maximum you requested, Network Firewall returns a NextToken value in the response. To retrieve the next batch of objects, use the token returned from the prior request in your next request.

$sel:firewalls:ListFirewallsResponse', listFirewallsResponse_firewalls - The firewall metadata objects for the VPCs that you specified. Depending on your setting for max results and the number of firewalls you have, a single call might not be the full list.

$sel:httpStatus:ListFirewallsResponse', listFirewallsResponse_httpStatus - The response's http status code.

Response Lenses

listFirewallsResponse_nextToken :: Lens' ListFirewallsResponse (Maybe Text)

When you request a list of objects with a MaxResults setting, if the number of objects that are still available for retrieval exceeds the maximum you requested, Network Firewall returns a NextToken value in the response. To retrieve the next batch of objects, use the token returned from the prior request in your next request.

listFirewallsResponse_firewalls :: Lens' ListFirewallsResponse (Maybe [FirewallMetadata])

The firewall metadata objects for the VPCs that you specified. Depending on your setting for max results and the number of firewalls you have, a single call might not be the full list.

listFirewallsResponse_httpStatus :: Lens' ListFirewallsResponse Int

The response's http status code.