Network.AWS.NetworkFirewall.Types.Firewall

Description

Documentation

data Firewall

The firewall defines the configuration settings for an AWS Network Firewall firewall. These settings include the firewall policy, the subnets in your VPC to use for the firewall endpoints, and any tags that are attached to the firewall AWS resource.

The status of the firewall, for example whether it's ready to filter network traffic, is provided in the corresponding FirewallStatus. You can retrieve both objects by calling DescribeFirewall.

See: newFirewall smart constructor.

Constructors

Firewall'

Fields

firewallArn :: Maybe Text

The Amazon Resource Name (ARN) of the firewall.

firewallPolicyChangeProtection :: Maybe Bool

A setting indicating whether the firewall is protected against a change to the firewall policy association. Use this setting to protect against accidentally modifying the firewall policy for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

subnetChangeProtection :: Maybe Bool

A setting indicating whether the firewall is protected against changes to the subnet associations. Use this setting to protect against accidentally modifying the subnet associations for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

deleteProtection :: Maybe Bool

A flag indicating whether it is possible to delete the firewall. A setting of TRUE indicates that the firewall is protected against deletion. Use this setting to protect against accidentally deleting a firewall that is in use. When you create a firewall, the operation initializes this flag to TRUE.

description :: Maybe Text

A description of the firewall.

tags :: Maybe (NonEmpty Tag)

firewallName :: Maybe Text

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

firewallPolicyArn :: Text

The Amazon Resource Name (ARN) of the firewall policy.

The relationship of firewall to firewall policy is many to one. Each firewall requires one firewall policy association, and you can use the same firewall policy for multiple firewalls.

vpcId :: Text

The unique identifier of the VPC where the firewall is in use.

subnetMappings :: [SubnetMapping]

The public subnets that Network Firewall is using for the firewall. Each subnet must belong to a different Availability Zone.

firewallId :: Text

The unique identifier for the firewall.

newFirewall

Arguments

:: Text
$sel:firewallPolicyArn:Firewall'

-> Text
$sel:vpcId:Firewall'

-> Text
$sel:firewallId:Firewall'

-> Firewall

Create a value of Firewall with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:firewallArn:Firewall', firewall_firewallArn - The Amazon Resource Name (ARN) of the firewall.

$sel:firewallPolicyChangeProtection:Firewall', firewall_firewallPolicyChangeProtection - A setting indicating whether the firewall is protected against a change to the firewall policy association. Use this setting to protect against accidentally modifying the firewall policy for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

$sel:subnetChangeProtection:Firewall', firewall_subnetChangeProtection - A setting indicating whether the firewall is protected against changes to the subnet associations. Use this setting to protect against accidentally modifying the subnet associations for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

$sel:deleteProtection:Firewall', firewall_deleteProtection - A flag indicating whether it is possible to delete the firewall. A setting of TRUE indicates that the firewall is protected against deletion. Use this setting to protect against accidentally deleting a firewall that is in use. When you create a firewall, the operation initializes this flag to TRUE.

$sel:description:Firewall', firewall_description - A description of the firewall.

$sel:tags:Firewall', firewall_tags -

$sel:firewallName:Firewall', firewall_firewallName - The descriptive name of the firewall. You can't change the name of a firewall after you create it.

$sel:firewallPolicyArn:Firewall', firewall_firewallPolicyArn - The Amazon Resource Name (ARN) of the firewall policy.

The relationship of firewall to firewall policy is many to one. Each firewall requires one firewall policy association, and you can use the same firewall policy for multiple firewalls.

$sel:vpcId:Firewall', firewall_vpcId - The unique identifier of the VPC where the firewall is in use.

$sel:subnetMappings:Firewall', firewall_subnetMappings - The public subnets that Network Firewall is using for the firewall. Each subnet must belong to a different Availability Zone.

$sel:firewallId:Firewall', firewall_firewallId - The unique identifier for the firewall.

firewall_firewallArn :: Lens' Firewall (Maybe Text)

The Amazon Resource Name (ARN) of the firewall.

firewall_firewallPolicyChangeProtection :: Lens' Firewall (Maybe Bool)

A setting indicating whether the firewall is protected against a change to the firewall policy association. Use this setting to protect against accidentally modifying the firewall policy for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

firewall_subnetChangeProtection :: Lens' Firewall (Maybe Bool)

A setting indicating whether the firewall is protected against changes to the subnet associations. Use this setting to protect against accidentally modifying the subnet associations for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

firewall_deleteProtection :: Lens' Firewall (Maybe Bool)

A flag indicating whether it is possible to delete the firewall. A setting of TRUE indicates that the firewall is protected against deletion. Use this setting to protect against accidentally deleting a firewall that is in use. When you create a firewall, the operation initializes this flag to TRUE.

firewall_description :: Lens' Firewall (Maybe Text)

A description of the firewall.

firewall_tags :: Lens' Firewall (Maybe (NonEmpty Tag))

firewall_firewallName :: Lens' Firewall (Maybe Text)

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

firewall_firewallPolicyArn :: Lens' Firewall Text

The Amazon Resource Name (ARN) of the firewall policy.

The relationship of firewall to firewall policy is many to one. Each firewall requires one firewall policy association, and you can use the same firewall policy for multiple firewalls.

firewall_vpcId :: Lens' Firewall Text

The unique identifier of the VPC where the firewall is in use.

firewall_subnetMappings :: Lens' Firewall [SubnetMapping]

The public subnets that Network Firewall is using for the firewall. Each subnet must belong to a different Availability Zone.

firewall_firewallId :: Lens' Firewall Text

The unique identifier for the firewall.