Network.AWS.NetworkFirewall.Types.LogDestinationConfig

Description

Documentation

data LogDestinationConfig

Defines where AWS Network Firewall sends logs for the firewall for one log type. This is used in LoggingConfiguration. You can send each type of log to an Amazon S3 bucket, a CloudWatch log group, or a Kinesis Data Firehose delivery stream.

Network Firewall generates logs for stateful rule groups. You can save alert and flow log types. The stateful rules engine records flow logs for all network traffic that it receives. It records alert logs for traffic that matches stateful rules that have the rule action set to DROP or ALERT.

See: newLogDestinationConfig smart constructor.

Constructors

LogDestinationConfig'

Fields

logType :: LogType

The type of log to send. Alert logs report traffic that matches a StatefulRule with an action setting that sends an alert log message. Flow logs are standard network traffic flow logs.

logDestinationType :: LogDestinationType

The type of storage destination to send these logs to. You can send logs to an Amazon S3 bucket, a CloudWatch log group, or a Kinesis Data Firehose delivery stream.

logDestination :: HashMap Text Text

The named location for the logs, provided in a key:value mapping that is specific to the chosen destination type.

For an Amazon S3 bucket, provide the name of the bucket, with key bucketName, and optionally provide a prefix, with key prefix. The following example specifies an Amazon S3 bucket named DOC-EXAMPLE-BUCKET and the prefix alerts:

"LogDestination": { "bucketName": "DOC-EXAMPLE-BUCKET", "prefix": "alerts" }

For a CloudWatch log group, provide the name of the CloudWatch log group, with key logGroup. The following example specifies a log group named alert-log-group:

"LogDestination": { "logGroup": "alert-log-group" }

For a Kinesis Data Firehose delivery stream, provide the name of the delivery stream, with key deliveryStream. The following example specifies a delivery stream named alert-delivery-stream:

"LogDestination": { "deliveryStream": "alert-delivery-stream" }

newLogDestinationConfig

Arguments

:: LogType
$sel:logType:LogDestinationConfig'

-> LogDestinationType
$sel:logDestinationType:LogDestinationConfig'

-> LogDestinationConfig

Create a value of LogDestinationConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logType:LogDestinationConfig', logDestinationConfig_logType - The type of log to send. Alert logs report traffic that matches a StatefulRule with an action setting that sends an alert log message. Flow logs are standard network traffic flow logs.

$sel:logDestinationType:LogDestinationConfig', logDestinationConfig_logDestinationType - The type of storage destination to send these logs to. You can send logs to an Amazon S3 bucket, a CloudWatch log group, or a Kinesis Data Firehose delivery stream.

$sel:logDestination:LogDestinationConfig', logDestinationConfig_logDestination - The named location for the logs, provided in a key:value mapping that is specific to the chosen destination type.

For an Amazon S3 bucket, provide the name of the bucket, with key bucketName, and optionally provide a prefix, with key prefix. The following example specifies an Amazon S3 bucket named DOC-EXAMPLE-BUCKET and the prefix alerts:

"LogDestination": { "bucketName": "DOC-EXAMPLE-BUCKET", "prefix": "alerts" }

For a CloudWatch log group, provide the name of the CloudWatch log group, with key logGroup. The following example specifies a log group named alert-log-group:

"LogDestination": { "logGroup": "alert-log-group" }

For a Kinesis Data Firehose delivery stream, provide the name of the delivery stream, with key deliveryStream. The following example specifies a delivery stream named alert-delivery-stream:

"LogDestination": { "deliveryStream": "alert-delivery-stream" }

logDestinationConfig_logType :: Lens' LogDestinationConfig LogType

The type of log to send. Alert logs report traffic that matches a StatefulRule with an action setting that sends an alert log message. Flow logs are standard network traffic flow logs.

logDestinationConfig_logDestinationType :: Lens' LogDestinationConfig LogDestinationType

The type of storage destination to send these logs to. You can send logs to an Amazon S3 bucket, a CloudWatch log group, or a Kinesis Data Firehose delivery stream.

logDestinationConfig_logDestination :: Lens' LogDestinationConfig (HashMap Text Text)

The named location for the logs, provided in a key:value mapping that is specific to the chosen destination type.

For an Amazon S3 bucket, provide the name of the bucket, with key bucketName, and optionally provide a prefix, with key prefix. The following example specifies an Amazon S3 bucket named DOC-EXAMPLE-BUCKET and the prefix alerts:

"LogDestination": { "bucketName": "DOC-EXAMPLE-BUCKET", "prefix": "alerts" }

For a CloudWatch log group, provide the name of the CloudWatch log group, with key logGroup. The following example specifies a log group named alert-log-group:

"LogDestination": { "logGroup": "alert-log-group" }

For a Kinesis Data Firehose delivery stream, provide the name of the delivery stream, with key deliveryStream. The following example specifies a delivery stream named alert-delivery-stream:

"LogDestination": { "deliveryStream": "alert-delivery-stream" }