Network.AWS.NetworkFirewall.Types.RuleGroupResponse

Description

Documentation

data RuleGroupResponse

The high-level properties of a rule group. This, along with the RuleGroup, define the rule group. You can retrieve all objects for a rule group by calling DescribeRuleGroup.

See: newRuleGroupResponse smart constructor.

Constructors

RuleGroupResponse'

Fields

numberOfAssociations :: Maybe Int

The number of firewall policies that use this rule group.

capacity :: Maybe Int

The maximum operating resources that this rule group can use. Rule group capacity is fixed at creation. When you update a rule group, you are limited to this capacity. When you reference a rule group from a firewall policy, Network Firewall reserves this capacity for the rule group.

You can retrieve the capacity that would be required for a rule group before you create the rule group by calling CreateRuleGroup with DryRun set to TRUE.

consumedCapacity :: Maybe Int

The number of capacity units currently consumed by the rule group rules.

ruleGroupStatus :: Maybe ResourceStatus

Detailed information about the current status of a rule group.

type' :: Maybe RuleGroupType

Indicates whether the rule group is stateless or stateful. If the rule group is stateless, it contains stateless rules. If it is stateful, it contains stateful rules.

description :: Maybe Text

A description of the rule group.

tags :: Maybe (NonEmpty Tag)

The key:value pairs to associate with the resource.

ruleGroupArn :: Text

The Amazon Resource Name (ARN) of the rule group.

If this response is for a create request that had DryRun set to TRUE, then this ARN is a placeholder that isn't attached to a valid resource.

ruleGroupName :: Text

The descriptive name of the rule group. You can't change the name of a rule group after you create it.

ruleGroupId :: Text

The unique identifier for the rule group.

newRuleGroupResponse

Arguments

:: Text
$sel:ruleGroupArn:RuleGroupResponse'

-> Text
$sel:ruleGroupName:RuleGroupResponse'

-> Text
$sel:ruleGroupId:RuleGroupResponse'

-> RuleGroupResponse

Create a value of RuleGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:numberOfAssociations:RuleGroupResponse', ruleGroupResponse_numberOfAssociations - The number of firewall policies that use this rule group.

$sel:capacity:RuleGroupResponse', ruleGroupResponse_capacity - The maximum operating resources that this rule group can use. Rule group capacity is fixed at creation. When you update a rule group, you are limited to this capacity. When you reference a rule group from a firewall policy, Network Firewall reserves this capacity for the rule group.

You can retrieve the capacity that would be required for a rule group before you create the rule group by calling CreateRuleGroup with DryRun set to TRUE.

$sel:consumedCapacity:RuleGroupResponse', ruleGroupResponse_consumedCapacity - The number of capacity units currently consumed by the rule group rules.

$sel:ruleGroupStatus:RuleGroupResponse', ruleGroupResponse_ruleGroupStatus - Detailed information about the current status of a rule group.

$sel:type':RuleGroupResponse', ruleGroupResponse_type - Indicates whether the rule group is stateless or stateful. If the rule group is stateless, it contains stateless rules. If it is stateful, it contains stateful rules.

$sel:description:RuleGroupResponse', ruleGroupResponse_description - A description of the rule group.

$sel:tags:RuleGroupResponse', ruleGroupResponse_tags - The key:value pairs to associate with the resource.

$sel:ruleGroupArn:RuleGroupResponse', ruleGroupResponse_ruleGroupArn - The Amazon Resource Name (ARN) of the rule group.

If this response is for a create request that had DryRun set to TRUE, then this ARN is a placeholder that isn't attached to a valid resource.

$sel:ruleGroupName:RuleGroupResponse', ruleGroupResponse_ruleGroupName - The descriptive name of the rule group. You can't change the name of a rule group after you create it.

$sel:ruleGroupId:RuleGroupResponse', ruleGroupResponse_ruleGroupId - The unique identifier for the rule group.

ruleGroupResponse_numberOfAssociations :: Lens' RuleGroupResponse (Maybe Int)

The number of firewall policies that use this rule group.

ruleGroupResponse_capacity :: Lens' RuleGroupResponse (Maybe Int)

The maximum operating resources that this rule group can use. Rule group capacity is fixed at creation. When you update a rule group, you are limited to this capacity. When you reference a rule group from a firewall policy, Network Firewall reserves this capacity for the rule group.

You can retrieve the capacity that would be required for a rule group before you create the rule group by calling CreateRuleGroup with DryRun set to TRUE.

ruleGroupResponse_consumedCapacity :: Lens' RuleGroupResponse (Maybe Int)

The number of capacity units currently consumed by the rule group rules.

ruleGroupResponse_ruleGroupStatus :: Lens' RuleGroupResponse (Maybe ResourceStatus)

Detailed information about the current status of a rule group.

ruleGroupResponse_type :: Lens' RuleGroupResponse (Maybe RuleGroupType)

Indicates whether the rule group is stateless or stateful. If the rule group is stateless, it contains stateless rules. If it is stateful, it contains stateful rules.

ruleGroupResponse_description :: Lens' RuleGroupResponse (Maybe Text)

A description of the rule group.

ruleGroupResponse_tags :: Lens' RuleGroupResponse (Maybe (NonEmpty Tag))

The key:value pairs to associate with the resource.

ruleGroupResponse_ruleGroupArn :: Lens' RuleGroupResponse Text

The Amazon Resource Name (ARN) of the rule group.

If this response is for a create request that had DryRun set to TRUE, then this ARN is a placeholder that isn't attached to a valid resource.

ruleGroupResponse_ruleGroupName :: Lens' RuleGroupResponse Text

The descriptive name of the rule group. You can't change the name of a rule group after you create it.

ruleGroupResponse_ruleGroupId :: Lens' RuleGroupResponse Text

The unique identifier for the rule group.