Network.AWS.NetworkFirewall.Types.RulesSource

Description

Documentation

data RulesSource

The stateless or stateful rules definitions for use in a single rule group. Each rule group requires a single RulesSource. You can use an instance of this for either stateless rules or stateful rules.

See: newRulesSource smart constructor.

Constructors

RulesSource'

Fields

rulesString :: Maybe Text

Stateful inspection criteria, provided in Suricata compatible intrusion prevention system (IPS) rules. Suricata is an open-source network IPS that includes a standard rule-based language for network traffic inspection.

These rules contain the inspection criteria and the action to take for traffic that matches the criteria, so this type of rule group doesn't have a separate action setting.

rulesSourceList :: Maybe RulesSourceList

Stateful inspection criteria for a domain list rule group.

statefulRules :: Maybe [StatefulRule]

An array of individual stateful rules inspection criteria to be used together in a stateful rule group. Use this option to specify simple Suricata rules with protocol, source and destination, ports, direction, and rule options. For information about the Suricata Rules format, see Rules Format.

statelessRulesAndCustomActions :: Maybe StatelessRulesAndCustomActions

Stateless inspection criteria to be used in a stateless rule group.

newRulesSource :: RulesSource

Create a value of RulesSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rulesString:RulesSource', rulesSource_rulesString - Stateful inspection criteria, provided in Suricata compatible intrusion prevention system (IPS) rules. Suricata is an open-source network IPS that includes a standard rule-based language for network traffic inspection.

These rules contain the inspection criteria and the action to take for traffic that matches the criteria, so this type of rule group doesn't have a separate action setting.

$sel:rulesSourceList:RulesSource', rulesSource_rulesSourceList - Stateful inspection criteria for a domain list rule group.

$sel:statefulRules:RulesSource', rulesSource_statefulRules - An array of individual stateful rules inspection criteria to be used together in a stateful rule group. Use this option to specify simple Suricata rules with protocol, source and destination, ports, direction, and rule options. For information about the Suricata Rules format, see Rules Format.

$sel:statelessRulesAndCustomActions:RulesSource', rulesSource_statelessRulesAndCustomActions - Stateless inspection criteria to be used in a stateless rule group.

rulesSource_rulesString :: Lens' RulesSource (Maybe Text)

Stateful inspection criteria, provided in Suricata compatible intrusion prevention system (IPS) rules. Suricata is an open-source network IPS that includes a standard rule-based language for network traffic inspection.

These rules contain the inspection criteria and the action to take for traffic that matches the criteria, so this type of rule group doesn't have a separate action setting.

rulesSource_rulesSourceList :: Lens' RulesSource (Maybe RulesSourceList)

Stateful inspection criteria for a domain list rule group.

rulesSource_statefulRules :: Lens' RulesSource (Maybe [StatefulRule])

An array of individual stateful rules inspection criteria to be used together in a stateful rule group. Use this option to specify simple Suricata rules with protocol, source and destination, ports, direction, and rule options. For information about the Suricata Rules format, see Rules Format.

rulesSource_statelessRulesAndCustomActions :: Lens' RulesSource (Maybe StatelessRulesAndCustomActions)

Stateless inspection criteria to be used in a stateless rule group.