Network.AWS.NetworkFirewall.Types.SubnetMapping

Description

Documentation

data SubnetMapping

The ID for a subnet that you want to associate with the firewall. This is used with CreateFirewall and AssociateSubnets. AWS Network Firewall creates an instance of the associated firewall in each subnet that you specify, to filter traffic in the subnet's Availability Zone.

See: newSubnetMapping smart constructor.

Constructors

SubnetMapping'

Fields

subnetId :: Text

The unique identifier for the subnet.

newSubnetMapping

Arguments

:: Text
$sel:subnetId:SubnetMapping'

-> SubnetMapping

Create a value of SubnetMapping with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subnetId:SubnetMapping', subnetMapping_subnetId - The unique identifier for the subnet.

subnetMapping_subnetId :: Lens' SubnetMapping Text

The unique identifier for the subnet.