Network.AWS.NetworkFirewall.UpdateRuleGroup

Description

Updates the rule settings for the specified rule group. You use a rule group by reference in one or more firewall policies. When you modify a rule group, you modify all firewall policies that use the rule group.

To update a rule group, first call DescribeRuleGroup to retrieve the current RuleGroup object, update the object as needed, and then provide the updated object to this call.

Creating a Request

data UpdateRuleGroup

See: newUpdateRuleGroup smart constructor.

Constructors

UpdateRuleGroup'

Fields

ruleGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of the rule group.

You must specify the ARN or the name, and you can specify both.

rules :: Maybe Text

A string containing stateful rule group rules specifications in Suricata flat format, with one rule per line. Use this to import your existing Suricata compatible rule groups.

You must provide either this rules setting or a populated RuleGroup setting, but not both.

You can provide your rule group specification in Suricata flat format through this setting when you create or update your rule group. The call response returns a RuleGroup object that Network Firewall has populated from your string.

type' :: Maybe RuleGroupType

Indicates whether the rule group is stateless or stateful. If the rule group is stateless, it contains stateless rules. If it is stateful, it contains stateful rules.

This setting is required for requests that do not include the RuleGroupARN.

description :: Maybe Text

A description of the rule group.

ruleGroupName :: Maybe Text

The descriptive name of the rule group. You can't change the name of a rule group after you create it.

You must specify the ARN or the name, and you can specify both.

dryRun :: Maybe Bool

Indicates whether you want Network Firewall to just check the validity of the request, rather than run the request.

If set to TRUE, Network Firewall checks whether the request can run successfully, but doesn't actually make the requested changes. The call returns the value that the request would return if you ran it with dry run set to FALSE, but doesn't make additions or changes to your resources. This option allows you to make sure that you have the required permissions to run the request and that your request parameters are valid.

If set to FALSE, Network Firewall makes the requested changes to your resources.

ruleGroup :: Maybe RuleGroup

An object that defines the rule group rules.

You must provide either this rule group setting or a Rules setting, but not both.

updateToken :: Text

A token used for optimistic locking. Network Firewall returns a token to your requests that access the rule group. The token marks the state of the rule group resource at the time of the request.

To make changes to the rule group, you provide the token in your request. Network Firewall uses the token to ensure that the rule group hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the rule group again to get a current copy of it with a current token. Reapply your changes as needed, then try the operation again using the new token.

newUpdateRuleGroup

Arguments

:: Text
$sel:updateToken:UpdateRuleGroup'

-> UpdateRuleGroup

Create a value of UpdateRuleGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ruleGroupArn:UpdateRuleGroup', updateRuleGroup_ruleGroupArn - The Amazon Resource Name (ARN) of the rule group.

You must specify the ARN or the name, and you can specify both.

$sel:rules:UpdateRuleGroup', updateRuleGroup_rules - A string containing stateful rule group rules specifications in Suricata flat format, with one rule per line. Use this to import your existing Suricata compatible rule groups.

You must provide either this rules setting or a populated RuleGroup setting, but not both.

You can provide your rule group specification in Suricata flat format through this setting when you create or update your rule group. The call response returns a RuleGroup object that Network Firewall has populated from your string.

$sel:type':UpdateRuleGroup', updateRuleGroup_type - Indicates whether the rule group is stateless or stateful. If the rule group is stateless, it contains stateless rules. If it is stateful, it contains stateful rules.

This setting is required for requests that do not include the RuleGroupARN.

$sel:description:UpdateRuleGroup', updateRuleGroup_description - A description of the rule group.

$sel:ruleGroupName:UpdateRuleGroup', updateRuleGroup_ruleGroupName - The descriptive name of the rule group. You can't change the name of a rule group after you create it.

You must specify the ARN or the name, and you can specify both.

$sel:dryRun:UpdateRuleGroup', updateRuleGroup_dryRun - Indicates whether you want Network Firewall to just check the validity of the request, rather than run the request.

If set to TRUE, Network Firewall checks whether the request can run successfully, but doesn't actually make the requested changes. The call returns the value that the request would return if you ran it with dry run set to FALSE, but doesn't make additions or changes to your resources. This option allows you to make sure that you have the required permissions to run the request and that your request parameters are valid.

If set to FALSE, Network Firewall makes the requested changes to your resources.

$sel:ruleGroup:UpdateRuleGroup', updateRuleGroup_ruleGroup - An object that defines the rule group rules.

You must provide either this rule group setting or a Rules setting, but not both.

$sel:updateToken:UpdateRuleGroup', updateRuleGroup_updateToken - A token used for optimistic locking. Network Firewall returns a token to your requests that access the rule group. The token marks the state of the rule group resource at the time of the request.

To make changes to the rule group, you provide the token in your request. Network Firewall uses the token to ensure that the rule group hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the rule group again to get a current copy of it with a current token. Reapply your changes as needed, then try the operation again using the new token.

Request Lenses

updateRuleGroup_ruleGroupArn :: Lens' UpdateRuleGroup (Maybe Text)

The Amazon Resource Name (ARN) of the rule group.

You must specify the ARN or the name, and you can specify both.

updateRuleGroup_rules :: Lens' UpdateRuleGroup (Maybe Text)

A string containing stateful rule group rules specifications in Suricata flat format, with one rule per line. Use this to import your existing Suricata compatible rule groups.

You must provide either this rules setting or a populated RuleGroup setting, but not both.

You can provide your rule group specification in Suricata flat format through this setting when you create or update your rule group. The call response returns a RuleGroup object that Network Firewall has populated from your string.

updateRuleGroup_type :: Lens' UpdateRuleGroup (Maybe RuleGroupType)

Indicates whether the rule group is stateless or stateful. If the rule group is stateless, it contains stateless rules. If it is stateful, it contains stateful rules.

This setting is required for requests that do not include the RuleGroupARN.

updateRuleGroup_description :: Lens' UpdateRuleGroup (Maybe Text)

A description of the rule group.

updateRuleGroup_ruleGroupName :: Lens' UpdateRuleGroup (Maybe Text)

The descriptive name of the rule group. You can't change the name of a rule group after you create it.

You must specify the ARN or the name, and you can specify both.

updateRuleGroup_dryRun :: Lens' UpdateRuleGroup (Maybe Bool)

Indicates whether you want Network Firewall to just check the validity of the request, rather than run the request.

If set to TRUE, Network Firewall checks whether the request can run successfully, but doesn't actually make the requested changes. The call returns the value that the request would return if you ran it with dry run set to FALSE, but doesn't make additions or changes to your resources. This option allows you to make sure that you have the required permissions to run the request and that your request parameters are valid.

If set to FALSE, Network Firewall makes the requested changes to your resources.

updateRuleGroup_ruleGroup :: Lens' UpdateRuleGroup (Maybe RuleGroup)

An object that defines the rule group rules.

You must provide either this rule group setting or a Rules setting, but not both.

updateRuleGroup_updateToken :: Lens' UpdateRuleGroup Text

A token used for optimistic locking. Network Firewall returns a token to your requests that access the rule group. The token marks the state of the rule group resource at the time of the request.

To make changes to the rule group, you provide the token in your request. Network Firewall uses the token to ensure that the rule group hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the rule group again to get a current copy of it with a current token. Reapply your changes as needed, then try the operation again using the new token.

Destructuring the Response

data UpdateRuleGroupResponse

See: newUpdateRuleGroupResponse smart constructor.

Constructors

UpdateRuleGroupResponse'

Fields

httpStatus :: Int

The response's http status code.

updateToken :: Text

A token used for optimistic locking. Network Firewall returns a token to your requests that access the rule group. The token marks the state of the rule group resource at the time of the request.

To make changes to the rule group, you provide the token in your request. Network Firewall uses the token to ensure that the rule group hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the rule group again to get a current copy of it with a current token. Reapply your changes as needed, then try the operation again using the new token.

ruleGroupResponse :: RuleGroupResponse

The high-level properties of a rule group. This, along with the RuleGroup, define the rule group. You can retrieve all objects for a rule group by calling DescribeRuleGroup.

newUpdateRuleGroupResponse

Arguments

:: Int
$sel:httpStatus:UpdateRuleGroupResponse'

-> Text
$sel:updateToken:UpdateRuleGroup'

-> RuleGroupResponse
$sel:ruleGroupResponse:UpdateRuleGroupResponse'

-> UpdateRuleGroupResponse

Create a value of UpdateRuleGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateRuleGroupResponse', updateRuleGroupResponse_httpStatus - The response's http status code.

$sel:updateToken:UpdateRuleGroup', updateRuleGroupResponse_updateToken - A token used for optimistic locking. Network Firewall returns a token to your requests that access the rule group. The token marks the state of the rule group resource at the time of the request.

To make changes to the rule group, you provide the token in your request. Network Firewall uses the token to ensure that the rule group hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the rule group again to get a current copy of it with a current token. Reapply your changes as needed, then try the operation again using the new token.

$sel:ruleGroupResponse:UpdateRuleGroupResponse', updateRuleGroupResponse_ruleGroupResponse - The high-level properties of a rule group. This, along with the RuleGroup, define the rule group. You can retrieve all objects for a rule group by calling DescribeRuleGroup.

Response Lenses

updateRuleGroupResponse_httpStatus :: Lens' UpdateRuleGroupResponse Int

The response's http status code.

updateRuleGroupResponse_updateToken :: Lens' UpdateRuleGroupResponse Text

A token used for optimistic locking. Network Firewall returns a token to your requests that access the rule group. The token marks the state of the rule group resource at the time of the request.

To make changes to the rule group, you provide the token in your request. Network Firewall uses the token to ensure that the rule group hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the rule group again to get a current copy of it with a current token. Reapply your changes as needed, then try the operation again using the new token.

updateRuleGroupResponse_ruleGroupResponse :: Lens' UpdateRuleGroupResponse RuleGroupResponse

The high-level properties of a rule group. This, along with the RuleGroup, define the rule group. You can retrieve all objects for a rule group by calling DescribeRuleGroup.