Network.AWS.NetworkFirewall.UpdateSubnetChangeProtection

Description

Creating a Request

data UpdateSubnetChangeProtection

See: newUpdateSubnetChangeProtection smart constructor.

Constructors

UpdateSubnetChangeProtection'

Fields

updateToken :: Maybe Text

An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

firewallArn :: Maybe Text

The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.

firewallName :: Maybe Text

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

You must specify the ARN or the name, and you can specify both.

subnetChangeProtection :: Bool

A setting indicating whether the firewall is protected against changes to the subnet associations. Use this setting to protect against accidentally modifying the subnet associations for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

newUpdateSubnetChangeProtection

Arguments

:: Bool
$sel:subnetChangeProtection:UpdateSubnetChangeProtection'

-> UpdateSubnetChangeProtection

Create a value of UpdateSubnetChangeProtection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateToken:UpdateSubnetChangeProtection', updateSubnetChangeProtection_updateToken - An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

$sel:firewallArn:UpdateSubnetChangeProtection', updateSubnetChangeProtection_firewallArn - The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.

$sel:firewallName:UpdateSubnetChangeProtection', updateSubnetChangeProtection_firewallName - The descriptive name of the firewall. You can't change the name of a firewall after you create it.

You must specify the ARN or the name, and you can specify both.

$sel:subnetChangeProtection:UpdateSubnetChangeProtection', updateSubnetChangeProtection_subnetChangeProtection - A setting indicating whether the firewall is protected against changes to the subnet associations. Use this setting to protect against accidentally modifying the subnet associations for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

Request Lenses

updateSubnetChangeProtection_updateToken :: Lens' UpdateSubnetChangeProtection (Maybe Text)

An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

updateSubnetChangeProtection_firewallArn :: Lens' UpdateSubnetChangeProtection (Maybe Text)

The Amazon Resource Name (ARN) of the firewall.

You must specify the ARN or the name, and you can specify both.

updateSubnetChangeProtection_firewallName :: Lens' UpdateSubnetChangeProtection (Maybe Text)

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

You must specify the ARN or the name, and you can specify both.

updateSubnetChangeProtection_subnetChangeProtection :: Lens' UpdateSubnetChangeProtection Bool

A setting indicating whether the firewall is protected against changes to the subnet associations. Use this setting to protect against accidentally modifying the subnet associations for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

Destructuring the Response

data UpdateSubnetChangeProtectionResponse

See: newUpdateSubnetChangeProtectionResponse smart constructor.

Constructors

UpdateSubnetChangeProtectionResponse'

Fields

updateToken :: Maybe Text

An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

firewallArn :: Maybe Text

The Amazon Resource Name (ARN) of the firewall.

subnetChangeProtection :: Maybe Bool

A setting indicating whether the firewall is protected against changes to the subnet associations. Use this setting to protect against accidentally modifying the subnet associations for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

firewallName :: Maybe Text

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

httpStatus :: Int

The response's http status code.

newUpdateSubnetChangeProtectionResponse

Arguments

:: Int
$sel:httpStatus:UpdateSubnetChangeProtectionResponse'

-> UpdateSubnetChangeProtectionResponse

Create a value of UpdateSubnetChangeProtectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateToken:UpdateSubnetChangeProtection', updateSubnetChangeProtectionResponse_updateToken - An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

$sel:firewallArn:UpdateSubnetChangeProtection', updateSubnetChangeProtectionResponse_firewallArn - The Amazon Resource Name (ARN) of the firewall.

$sel:subnetChangeProtection:UpdateSubnetChangeProtection', updateSubnetChangeProtectionResponse_subnetChangeProtection - A setting indicating whether the firewall is protected against changes to the subnet associations. Use this setting to protect against accidentally modifying the subnet associations for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

$sel:firewallName:UpdateSubnetChangeProtection', updateSubnetChangeProtectionResponse_firewallName - The descriptive name of the firewall. You can't change the name of a firewall after you create it.

$sel:httpStatus:UpdateSubnetChangeProtectionResponse', updateSubnetChangeProtectionResponse_httpStatus - The response's http status code.

Response Lenses

updateSubnetChangeProtectionResponse_updateToken :: Lens' UpdateSubnetChangeProtectionResponse (Maybe Text)

An optional token that you can use for optimistic locking. Network Firewall returns a token to your requests that access the firewall. The token marks the state of the firewall resource at the time of the request.

To make an unconditional change to the firewall, omit the token in your update request. Without the token, Network Firewall performs your updates regardless of whether the firewall has changed since you last retrieved it.

To make a conditional change to the firewall, provide the token in your update request. Network Firewall uses the token to ensure that the firewall hasn't changed since you last retrieved it. If it has changed, the operation fails with an InvalidTokenException. If this happens, retrieve the firewall again to get a current copy of it with a new token. Reapply your changes as needed, then try the operation again using the new token.

updateSubnetChangeProtectionResponse_firewallArn :: Lens' UpdateSubnetChangeProtectionResponse (Maybe Text)

The Amazon Resource Name (ARN) of the firewall.

updateSubnetChangeProtectionResponse_subnetChangeProtection :: Lens' UpdateSubnetChangeProtectionResponse (Maybe Bool)

A setting indicating whether the firewall is protected against changes to the subnet associations. Use this setting to protect against accidentally modifying the subnet associations for a firewall that is in use. When you create a firewall, the operation initializes this setting to TRUE.

updateSubnetChangeProtectionResponse_firewallName :: Lens' UpdateSubnetChangeProtectionResponse (Maybe Text)

The descriptive name of the firewall. You can't change the name of a firewall after you create it.

updateSubnetChangeProtectionResponse_httpStatus :: Lens' UpdateSubnetChangeProtectionResponse Int

The response's http status code.