Network.AWS.NetworkManager.CreateConnection

Description

Creates a connection between two devices. The devices can be a physical or virtual appliance that connects to a third-party appliance in a VPC, or a physical appliance that connects to another physical appliance in an on-premises network.

Creating a Request

data CreateConnection

See: newCreateConnection smart constructor.

Constructors

CreateConnection'

Fields

connectedLinkId :: Maybe Text

The ID of the link for the second device.

linkId :: Maybe Text

The ID of the link for the first device.

description :: Maybe Text

A description of the connection.

Length Constraints: Maximum length of 256 characters.

tags :: Maybe [Tag]

The tags to apply to the resource during creation.

globalNetworkId :: Text

The ID of the global network.

deviceId :: Text

The ID of the first device in the connection.

connectedDeviceId :: Text

The ID of the second device in the connection.

newCreateConnection

Arguments

:: Text
$sel:globalNetworkId:CreateConnection'

-> Text
$sel:deviceId:CreateConnection'

-> Text
$sel:connectedDeviceId:CreateConnection'

-> CreateConnection

Create a value of CreateConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectedLinkId:CreateConnection', createConnection_connectedLinkId - The ID of the link for the second device.

$sel:linkId:CreateConnection', createConnection_linkId - The ID of the link for the first device.

$sel:description:CreateConnection', createConnection_description - A description of the connection.

Length Constraints: Maximum length of 256 characters.

$sel:tags:CreateConnection', createConnection_tags - The tags to apply to the resource during creation.

$sel:globalNetworkId:CreateConnection', createConnection_globalNetworkId - The ID of the global network.

$sel:deviceId:CreateConnection', createConnection_deviceId - The ID of the first device in the connection.

$sel:connectedDeviceId:CreateConnection', createConnection_connectedDeviceId - The ID of the second device in the connection.

Request Lenses

createConnection_connectedLinkId :: Lens' CreateConnection (Maybe Text)

The ID of the link for the second device.

createConnection_linkId :: Lens' CreateConnection (Maybe Text)

The ID of the link for the first device.

createConnection_description :: Lens' CreateConnection (Maybe Text)

A description of the connection.

Length Constraints: Maximum length of 256 characters.

createConnection_tags :: Lens' CreateConnection (Maybe [Tag])

The tags to apply to the resource during creation.

createConnection_globalNetworkId :: Lens' CreateConnection Text

The ID of the global network.

createConnection_deviceId :: Lens' CreateConnection Text

The ID of the first device in the connection.

createConnection_connectedDeviceId :: Lens' CreateConnection Text

The ID of the second device in the connection.

Destructuring the Response

data CreateConnectionResponse

See: newCreateConnectionResponse smart constructor.

Constructors

CreateConnectionResponse'

Fields

connection :: Maybe Connection

Information about the connection.

httpStatus :: Int

The response's http status code.

newCreateConnectionResponse

Arguments

:: Int
$sel:httpStatus:CreateConnectionResponse'

-> CreateConnectionResponse

Create a value of CreateConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connection:CreateConnectionResponse', createConnectionResponse_connection - Information about the connection.

$sel:httpStatus:CreateConnectionResponse', createConnectionResponse_httpStatus - The response's http status code.

Response Lenses

createConnectionResponse_connection :: Lens' CreateConnectionResponse (Maybe Connection)

Information about the connection.

createConnectionResponse_httpStatus :: Lens' CreateConnectionResponse Int

The response's http status code.