Network.AWS.NetworkManager.CreateDevice

Description

Creates a new device in a global network. If you specify both a site ID and a location, the location of the site is used for visualization in the Network Manager console.

Creating a Request

data CreateDevice

See: newCreateDevice smart constructor.

Constructors

CreateDevice'

Fields

vendor :: Maybe Text

The vendor of the device.

Length Constraints: Maximum length of 128 characters.

location :: Maybe (Sensitive Location)

The location of the device.

aWSLocation :: Maybe AWSLocation

The AWS location of the device.

model :: Maybe Text

The model of the device.

Length Constraints: Maximum length of 128 characters.

type' :: Maybe Text

The type of the device.

serialNumber :: Maybe Text

The serial number of the device.

Length Constraints: Maximum length of 128 characters.

siteId :: Maybe Text

The ID of the site.

description :: Maybe Text

A description of the device.

Length Constraints: Maximum length of 256 characters.

tags :: Maybe [Tag]

The tags to apply to the resource during creation.

globalNetworkId :: Text

The ID of the global network.

newCreateDevice

Arguments

:: Text
$sel:globalNetworkId:CreateDevice'

-> CreateDevice

Create a value of CreateDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vendor:CreateDevice', createDevice_vendor - The vendor of the device.

Length Constraints: Maximum length of 128 characters.

$sel:location:CreateDevice', createDevice_location - The location of the device.

$sel:aWSLocation:CreateDevice', createDevice_aWSLocation - The AWS location of the device.

$sel:model:CreateDevice', createDevice_model - The model of the device.

Length Constraints: Maximum length of 128 characters.

$sel:type':CreateDevice', createDevice_type - The type of the device.

$sel:serialNumber:CreateDevice', createDevice_serialNumber - The serial number of the device.

Length Constraints: Maximum length of 128 characters.

$sel:siteId:CreateDevice', createDevice_siteId - The ID of the site.

$sel:description:CreateDevice', createDevice_description - A description of the device.

Length Constraints: Maximum length of 256 characters.

$sel:tags:CreateDevice', createDevice_tags - The tags to apply to the resource during creation.

$sel:globalNetworkId:CreateDevice', createDevice_globalNetworkId - The ID of the global network.

Request Lenses

createDevice_vendor :: Lens' CreateDevice (Maybe Text)

The vendor of the device.

Length Constraints: Maximum length of 128 characters.

createDevice_location :: Lens' CreateDevice (Maybe Location)

The location of the device.

createDevice_aWSLocation :: Lens' CreateDevice (Maybe AWSLocation)

The AWS location of the device.

createDevice_model :: Lens' CreateDevice (Maybe Text)

The model of the device.

Length Constraints: Maximum length of 128 characters.

createDevice_type :: Lens' CreateDevice (Maybe Text)

The type of the device.

createDevice_serialNumber :: Lens' CreateDevice (Maybe Text)

The serial number of the device.

Length Constraints: Maximum length of 128 characters.

createDevice_siteId :: Lens' CreateDevice (Maybe Text)

The ID of the site.

createDevice_description :: Lens' CreateDevice (Maybe Text)

A description of the device.

Length Constraints: Maximum length of 256 characters.

createDevice_tags :: Lens' CreateDevice (Maybe [Tag])

The tags to apply to the resource during creation.

createDevice_globalNetworkId :: Lens' CreateDevice Text

The ID of the global network.

Destructuring the Response

data CreateDeviceResponse

See: newCreateDeviceResponse smart constructor.

Constructors

CreateDeviceResponse'

Fields

device :: Maybe Device

Information about the device.

httpStatus :: Int

The response's http status code.

newCreateDeviceResponse

Arguments

:: Int
$sel:httpStatus:CreateDeviceResponse'

-> CreateDeviceResponse

Create a value of CreateDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:device:CreateDeviceResponse', createDeviceResponse_device - Information about the device.

$sel:httpStatus:CreateDeviceResponse', createDeviceResponse_httpStatus - The response's http status code.

Response Lenses

createDeviceResponse_device :: Lens' CreateDeviceResponse (Maybe Device)

Information about the device.

createDeviceResponse_httpStatus :: Lens' CreateDeviceResponse Int

The response's http status code.