Network.AWS.NetworkManager.CreateGlobalNetwork

Description

Creates a new, empty global network.

Creating a Request

data CreateGlobalNetwork

See: newCreateGlobalNetwork smart constructor.

Constructors

CreateGlobalNetwork'

Fields

description :: Maybe Text

A description of the global network.

Length Constraints: Maximum length of 256 characters.

tags :: Maybe [Tag]

The tags to apply to the resource during creation.

newCreateGlobalNetwork :: CreateGlobalNetwork

Create a value of CreateGlobalNetwork with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateGlobalNetwork', createGlobalNetwork_description - A description of the global network.

Length Constraints: Maximum length of 256 characters.

$sel:tags:CreateGlobalNetwork', createGlobalNetwork_tags - The tags to apply to the resource during creation.

Request Lenses

createGlobalNetwork_description :: Lens' CreateGlobalNetwork (Maybe Text)

A description of the global network.

Length Constraints: Maximum length of 256 characters.

createGlobalNetwork_tags :: Lens' CreateGlobalNetwork (Maybe [Tag])

The tags to apply to the resource during creation.

Destructuring the Response

data CreateGlobalNetworkResponse

See: newCreateGlobalNetworkResponse smart constructor.

Constructors

CreateGlobalNetworkResponse'

Fields

globalNetwork :: Maybe GlobalNetwork

Information about the global network object.

httpStatus :: Int

The response's http status code.

newCreateGlobalNetworkResponse

Arguments

:: Int
$sel:httpStatus:CreateGlobalNetworkResponse'

-> CreateGlobalNetworkResponse

Create a value of CreateGlobalNetworkResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalNetwork:CreateGlobalNetworkResponse', createGlobalNetworkResponse_globalNetwork - Information about the global network object.

$sel:httpStatus:CreateGlobalNetworkResponse', createGlobalNetworkResponse_httpStatus - The response's http status code.

Response Lenses

createGlobalNetworkResponse_globalNetwork :: Lens' CreateGlobalNetworkResponse (Maybe GlobalNetwork)

Information about the global network object.

createGlobalNetworkResponse_httpStatus :: Lens' CreateGlobalNetworkResponse Int

The response's http status code.