Network.AWS.NetworkManager.DescribeGlobalNetworks

Description

Describes one or more global networks. By default, all global networks are described. To describe the objects in your global network, you must use the appropriate Get* action. For example, to list the transit gateways in your global network, use GetTransitGatewayRegistrations.

This operation returns paginated results.

Creating a Request

data DescribeGlobalNetworks

See: newDescribeGlobalNetworks smart constructor.

Constructors

DescribeGlobalNetworks'

Fields

nextToken :: Maybe Text

The token for the next page of results.

globalNetworkIds :: Maybe [Text]

The IDs of one or more global networks. The maximum is 10.

maxResults :: Maybe Natural

The maximum number of results to return.

newDescribeGlobalNetworks :: DescribeGlobalNetworks

Create a value of DescribeGlobalNetworks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeGlobalNetworks', describeGlobalNetworks_nextToken - The token for the next page of results.

$sel:globalNetworkIds:DescribeGlobalNetworks', describeGlobalNetworks_globalNetworkIds - The IDs of one or more global networks. The maximum is 10.

$sel:maxResults:DescribeGlobalNetworks', describeGlobalNetworks_maxResults - The maximum number of results to return.

Request Lenses

describeGlobalNetworks_nextToken :: Lens' DescribeGlobalNetworks (Maybe Text)

The token for the next page of results.

describeGlobalNetworks_globalNetworkIds :: Lens' DescribeGlobalNetworks (Maybe [Text])

The IDs of one or more global networks. The maximum is 10.

describeGlobalNetworks_maxResults :: Lens' DescribeGlobalNetworks (Maybe Natural)

The maximum number of results to return.

Destructuring the Response

data DescribeGlobalNetworksResponse

See: newDescribeGlobalNetworksResponse smart constructor.

Constructors

DescribeGlobalNetworksResponse'

Fields

globalNetworks :: Maybe [GlobalNetwork]

Information about the global networks.

nextToken :: Maybe Text

The token for the next page of results.

httpStatus :: Int

The response's http status code.

newDescribeGlobalNetworksResponse

Arguments

:: Int
$sel:httpStatus:DescribeGlobalNetworksResponse'

-> DescribeGlobalNetworksResponse

Create a value of DescribeGlobalNetworksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalNetworks:DescribeGlobalNetworksResponse', describeGlobalNetworksResponse_globalNetworks - Information about the global networks.

$sel:nextToken:DescribeGlobalNetworks', describeGlobalNetworksResponse_nextToken - The token for the next page of results.

$sel:httpStatus:DescribeGlobalNetworksResponse', describeGlobalNetworksResponse_httpStatus - The response's http status code.

Response Lenses

describeGlobalNetworksResponse_globalNetworks :: Lens' DescribeGlobalNetworksResponse (Maybe [GlobalNetwork])

Information about the global networks.

describeGlobalNetworksResponse_nextToken :: Lens' DescribeGlobalNetworksResponse (Maybe Text)

The token for the next page of results.

describeGlobalNetworksResponse_httpStatus :: Lens' DescribeGlobalNetworksResponse Int

The response's http status code.