Network.AWS.NetworkManager.Lens

Description

Operations

GetLinkAssociations

getLinkAssociations_nextToken :: Lens' GetLinkAssociations (Maybe Text)

The token for the next page of results.

getLinkAssociations_linkId :: Lens' GetLinkAssociations (Maybe Text)

The ID of the link.

getLinkAssociations_deviceId :: Lens' GetLinkAssociations (Maybe Text)

The ID of the device.

getLinkAssociations_maxResults :: Lens' GetLinkAssociations (Maybe Natural)

The maximum number of results to return.

getLinkAssociations_globalNetworkId :: Lens' GetLinkAssociations Text

The ID of the global network.

getLinkAssociationsResponse_nextToken :: Lens' GetLinkAssociationsResponse (Maybe Text)

The token for the next page of results.

getLinkAssociationsResponse_linkAssociations :: Lens' GetLinkAssociationsResponse (Maybe [LinkAssociation])

The link associations.

getLinkAssociationsResponse_httpStatus :: Lens' GetLinkAssociationsResponse Int

The response's http status code.

AssociateLink

associateLink_globalNetworkId :: Lens' AssociateLink Text

The ID of the global network.

associateLink_deviceId :: Lens' AssociateLink Text

The ID of the device.

associateLink_linkId :: Lens' AssociateLink Text

The ID of the link.

associateLinkResponse_linkAssociation :: Lens' AssociateLinkResponse (Maybe LinkAssociation)

The link association.

associateLinkResponse_httpStatus :: Lens' AssociateLinkResponse Int

The response's http status code.

AssociateTransitGatewayConnectPeer

associateTransitGatewayConnectPeer_linkId :: Lens' AssociateTransitGatewayConnectPeer (Maybe Text)

The ID of the link.

associateTransitGatewayConnectPeer_globalNetworkId :: Lens' AssociateTransitGatewayConnectPeer Text

The ID of the global network.

associateTransitGatewayConnectPeer_transitGatewayConnectPeerArn :: Lens' AssociateTransitGatewayConnectPeer Text

The Amazon Resource Name (ARN) of the Connect peer.

associateTransitGatewayConnectPeer_deviceId :: Lens' AssociateTransitGatewayConnectPeer Text

The ID of the device.

associateTransitGatewayConnectPeerResponse_transitGatewayConnectPeerAssociation :: Lens' AssociateTransitGatewayConnectPeerResponse (Maybe TransitGatewayConnectPeerAssociation)

The transit gateway Connect peer association.

associateTransitGatewayConnectPeerResponse_httpStatus :: Lens' AssociateTransitGatewayConnectPeerResponse Int

The response's http status code.

CreateSite

createSite_location :: Lens' CreateSite (Maybe Location)

The site location. This information is used for visualization in the Network Manager console. If you specify the address, the latitude and longitude are automatically calculated.

Address: The physical address of the site.

Latitude: The latitude of the site.

Longitude: The longitude of the site.

createSite_description :: Lens' CreateSite (Maybe Text)

A description of your site.

Length Constraints: Maximum length of 256 characters.

createSite_tags :: Lens' CreateSite (Maybe [Tag])

The tags to apply to the resource during creation.

createSite_globalNetworkId :: Lens' CreateSite Text

The ID of the global network.

createSiteResponse_site :: Lens' CreateSiteResponse (Maybe Site)

Information about the site.

createSiteResponse_httpStatus :: Lens' CreateSiteResponse Int

The response's http status code.

DeleteConnection

deleteConnection_globalNetworkId :: Lens' DeleteConnection Text

The ID of the global network.

deleteConnection_connectionId :: Lens' DeleteConnection Text

The ID of the connection.

deleteConnectionResponse_connection :: Lens' DeleteConnectionResponse (Maybe Connection)

Information about the connection.

deleteConnectionResponse_httpStatus :: Lens' DeleteConnectionResponse Int

The response's http status code.

UpdateConnection

updateConnection_connectedLinkId :: Lens' UpdateConnection (Maybe Text)

The ID of the link for the second device in the connection.

updateConnection_linkId :: Lens' UpdateConnection (Maybe Text)

The ID of the link for the first device in the connection.

updateConnection_description :: Lens' UpdateConnection (Maybe Text)

A description of the connection.

Length Constraints: Maximum length of 256 characters.

updateConnection_globalNetworkId :: Lens' UpdateConnection Text

The ID of the global network.

updateConnection_connectionId :: Lens' UpdateConnection Text

The ID of the connection.

updateConnectionResponse_connection :: Lens' UpdateConnectionResponse (Maybe Connection)

Information about the connection.

updateConnectionResponse_httpStatus :: Lens' UpdateConnectionResponse Int

The response's http status code.

DeregisterTransitGateway

deregisterTransitGateway_globalNetworkId :: Lens' DeregisterTransitGateway Text

The ID of the global network.

deregisterTransitGateway_transitGatewayArn :: Lens' DeregisterTransitGateway Text

The Amazon Resource Name (ARN) of the transit gateway.

deregisterTransitGatewayResponse_transitGatewayRegistration :: Lens' DeregisterTransitGatewayResponse (Maybe TransitGatewayRegistration)

The transit gateway registration information.

deregisterTransitGatewayResponse_httpStatus :: Lens' DeregisterTransitGatewayResponse Int

The response's http status code.

GetTransitGatewayConnectPeerAssociations

getTransitGatewayConnectPeerAssociations_transitGatewayConnectPeerArns :: Lens' GetTransitGatewayConnectPeerAssociations (Maybe [Text])

One or more transit gateway Connect peer Amazon Resource Names (ARNs).

getTransitGatewayConnectPeerAssociations_nextToken :: Lens' GetTransitGatewayConnectPeerAssociations (Maybe Text)

The token for the next page of results.

getTransitGatewayConnectPeerAssociations_maxResults :: Lens' GetTransitGatewayConnectPeerAssociations (Maybe Natural)

The maximum number of results to return.

getTransitGatewayConnectPeerAssociations_globalNetworkId :: Lens' GetTransitGatewayConnectPeerAssociations Text

The ID of the global network.

getTransitGatewayConnectPeerAssociationsResponse_nextToken :: Lens' GetTransitGatewayConnectPeerAssociationsResponse (Maybe Text)

The token to use for the next page of results.

getTransitGatewayConnectPeerAssociationsResponse_transitGatewayConnectPeerAssociations :: Lens' GetTransitGatewayConnectPeerAssociationsResponse (Maybe [TransitGatewayConnectPeerAssociation])

Information about the transit gateway Connect peer associations.

getTransitGatewayConnectPeerAssociationsResponse_httpStatus :: Lens' GetTransitGatewayConnectPeerAssociationsResponse Int

The response's http status code.

UpdateSite

updateSite_location :: Lens' UpdateSite (Maybe Location)

The site location:

Address: The physical address of the site.

Latitude: The latitude of the site.

Longitude: The longitude of the site.

updateSite_description :: Lens' UpdateSite (Maybe Text)

A description of your site.

Length Constraints: Maximum length of 256 characters.

updateSite_globalNetworkId :: Lens' UpdateSite Text

The ID of the global network.

updateSite_siteId :: Lens' UpdateSite Text

The ID of your site.

updateSiteResponse_site :: Lens' UpdateSiteResponse (Maybe Site)

Information about the site.

updateSiteResponse_httpStatus :: Lens' UpdateSiteResponse Int

The response's http status code.

DeleteSite

deleteSite_globalNetworkId :: Lens' DeleteSite Text

The ID of the global network.

deleteSite_siteId :: Lens' DeleteSite Text

The ID of the site.

deleteSiteResponse_site :: Lens' DeleteSiteResponse (Maybe Site)

Information about the site.

deleteSiteResponse_httpStatus :: Lens' DeleteSiteResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the resource.

listTagsForResourceResponse_tagList :: Lens' ListTagsForResourceResponse (Maybe [Tag])

The list of tags.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

DisassociateLink

disassociateLink_globalNetworkId :: Lens' DisassociateLink Text

The ID of the global network.

disassociateLink_deviceId :: Lens' DisassociateLink Text

The ID of the device.

disassociateLink_linkId :: Lens' DisassociateLink Text

The ID of the link.

disassociateLinkResponse_linkAssociation :: Lens' DisassociateLinkResponse (Maybe LinkAssociation)

Information about the link association.

disassociateLinkResponse_httpStatus :: Lens' DisassociateLinkResponse Int

The response's http status code.

CreateConnection

createConnection_connectedLinkId :: Lens' CreateConnection (Maybe Text)

The ID of the link for the second device.

createConnection_linkId :: Lens' CreateConnection (Maybe Text)

The ID of the link for the first device.

createConnection_description :: Lens' CreateConnection (Maybe Text)

A description of the connection.

Length Constraints: Maximum length of 256 characters.

createConnection_tags :: Lens' CreateConnection (Maybe [Tag])

The tags to apply to the resource during creation.

createConnection_globalNetworkId :: Lens' CreateConnection Text

The ID of the global network.

createConnection_deviceId :: Lens' CreateConnection Text

The ID of the first device in the connection.

createConnection_connectedDeviceId :: Lens' CreateConnection Text

The ID of the second device in the connection.

createConnectionResponse_connection :: Lens' CreateConnectionResponse (Maybe Connection)

Information about the connection.

createConnectionResponse_httpStatus :: Lens' CreateConnectionResponse Int

The response's http status code.

GetDevices

getDevices_deviceIds :: Lens' GetDevices (Maybe [Text])

One or more device IDs. The maximum is 10.

getDevices_nextToken :: Lens' GetDevices (Maybe Text)

The token for the next page of results.

getDevices_siteId :: Lens' GetDevices (Maybe Text)

The ID of the site.

getDevices_maxResults :: Lens' GetDevices (Maybe Natural)

The maximum number of results to return.

getDevices_globalNetworkId :: Lens' GetDevices Text

The ID of the global network.

getDevicesResponse_nextToken :: Lens' GetDevicesResponse (Maybe Text)

The token for the next page of results.

getDevicesResponse_devices :: Lens' GetDevicesResponse (Maybe [Device])

The devices.

getDevicesResponse_httpStatus :: Lens' GetDevicesResponse Int

The response's http status code.

GetLinks

getLinks_linkIds :: Lens' GetLinks (Maybe [Text])

One or more link IDs. The maximum is 10.

getLinks_nextToken :: Lens' GetLinks (Maybe Text)

The token for the next page of results.

getLinks_type :: Lens' GetLinks (Maybe Text)

The link type.

getLinks_siteId :: Lens' GetLinks (Maybe Text)

The ID of the site.

getLinks_maxResults :: Lens' GetLinks (Maybe Natural)

The maximum number of results to return.

getLinks_provider :: Lens' GetLinks (Maybe Text)

The link provider.

getLinks_globalNetworkId :: Lens' GetLinks Text

The ID of the global network.

getLinksResponse_nextToken :: Lens' GetLinksResponse (Maybe Text)

The token for the next page of results.

getLinksResponse_links :: Lens' GetLinksResponse (Maybe [Link])

The links.

getLinksResponse_httpStatus :: Lens' GetLinksResponse Int

The response's http status code.

DescribeGlobalNetworks

describeGlobalNetworks_nextToken :: Lens' DescribeGlobalNetworks (Maybe Text)

The token for the next page of results.

describeGlobalNetworks_globalNetworkIds :: Lens' DescribeGlobalNetworks (Maybe [Text])

The IDs of one or more global networks. The maximum is 10.

describeGlobalNetworks_maxResults :: Lens' DescribeGlobalNetworks (Maybe Natural)

The maximum number of results to return.

describeGlobalNetworksResponse_globalNetworks :: Lens' DescribeGlobalNetworksResponse (Maybe [GlobalNetwork])

Information about the global networks.

describeGlobalNetworksResponse_nextToken :: Lens' DescribeGlobalNetworksResponse (Maybe Text)

The token for the next page of results.

describeGlobalNetworksResponse_httpStatus :: Lens' DescribeGlobalNetworksResponse Int

The response's http status code.

DisassociateCustomerGateway

disassociateCustomerGateway_globalNetworkId :: Lens' DisassociateCustomerGateway Text

The ID of the global network.

disassociateCustomerGateway_customerGatewayArn :: Lens' DisassociateCustomerGateway Text

The Amazon Resource Name (ARN) of the customer gateway. For more information, see Resources Defined by Amazon EC2.

disassociateCustomerGatewayResponse_customerGatewayAssociation :: Lens' DisassociateCustomerGatewayResponse (Maybe CustomerGatewayAssociation)

Information about the customer gateway association.

disassociateCustomerGatewayResponse_httpStatus :: Lens' DisassociateCustomerGatewayResponse Int

The response's http status code.

DisassociateTransitGatewayConnectPeer

disassociateTransitGatewayConnectPeer_globalNetworkId :: Lens' DisassociateTransitGatewayConnectPeer Text

The ID of the global network.

disassociateTransitGatewayConnectPeer_transitGatewayConnectPeerArn :: Lens' DisassociateTransitGatewayConnectPeer Text

The Amazon Resource Name (ARN) of the transit gateway Connect peer.

disassociateTransitGatewayConnectPeerResponse_transitGatewayConnectPeerAssociation :: Lens' DisassociateTransitGatewayConnectPeerResponse (Maybe TransitGatewayConnectPeerAssociation)

The transit gateway Connect peer association.

disassociateTransitGatewayConnectPeerResponse_httpStatus :: Lens' DisassociateTransitGatewayConnectPeerResponse Int

The response's http status code.

CreateGlobalNetwork

createGlobalNetwork_description :: Lens' CreateGlobalNetwork (Maybe Text)

A description of the global network.

Length Constraints: Maximum length of 256 characters.

createGlobalNetwork_tags :: Lens' CreateGlobalNetwork (Maybe [Tag])

The tags to apply to the resource during creation.

createGlobalNetworkResponse_globalNetwork :: Lens' CreateGlobalNetworkResponse (Maybe GlobalNetwork)

Information about the global network object.

createGlobalNetworkResponse_httpStatus :: Lens' CreateGlobalNetworkResponse Int

The response's http status code.

CreateLink

createLink_type :: Lens' CreateLink (Maybe Text)

The type of the link.

Constraints: Cannot include the following characters: | \ ^

Length Constraints: Maximum length of 128 characters.

createLink_description :: Lens' CreateLink (Maybe Text)

A description of the link.

Length Constraints: Maximum length of 256 characters.

createLink_provider :: Lens' CreateLink (Maybe Text)

The provider of the link.

Constraints: Cannot include the following characters: | \ ^

Length Constraints: Maximum length of 128 characters.

createLink_tags :: Lens' CreateLink (Maybe [Tag])

The tags to apply to the resource during creation.

createLink_globalNetworkId :: Lens' CreateLink Text

The ID of the global network.

createLink_bandwidth :: Lens' CreateLink Bandwidth

The upload speed and download speed in Mbps.

createLink_siteId :: Lens' CreateLink Text

The ID of the site.

createLinkResponse_link :: Lens' CreateLinkResponse (Maybe Link)

Information about the link.

createLinkResponse_httpStatus :: Lens' CreateLinkResponse Int

The response's http status code.

DeleteGlobalNetwork

deleteGlobalNetwork_globalNetworkId :: Lens' DeleteGlobalNetwork Text

The ID of the global network.

deleteGlobalNetworkResponse_globalNetwork :: Lens' DeleteGlobalNetworkResponse (Maybe GlobalNetwork)

Information about the global network.

deleteGlobalNetworkResponse_httpStatus :: Lens' DeleteGlobalNetworkResponse Int

The response's http status code.

UpdateGlobalNetwork

updateGlobalNetwork_description :: Lens' UpdateGlobalNetwork (Maybe Text)

A description of the global network.

Length Constraints: Maximum length of 256 characters.

updateGlobalNetwork_globalNetworkId :: Lens' UpdateGlobalNetwork Text

The ID of your global network.

updateGlobalNetworkResponse_globalNetwork :: Lens' UpdateGlobalNetworkResponse (Maybe GlobalNetwork)

Information about the global network object.

updateGlobalNetworkResponse_httpStatus :: Lens' UpdateGlobalNetworkResponse Int

The response's http status code.

CreateDevice

createDevice_vendor :: Lens' CreateDevice (Maybe Text)

The vendor of the device.

Length Constraints: Maximum length of 128 characters.

createDevice_location :: Lens' CreateDevice (Maybe Location)

The location of the device.

createDevice_aWSLocation :: Lens' CreateDevice (Maybe AWSLocation)

The AWS location of the device.

createDevice_model :: Lens' CreateDevice (Maybe Text)

The model of the device.

Length Constraints: Maximum length of 128 characters.

createDevice_type :: Lens' CreateDevice (Maybe Text)

The type of the device.

createDevice_serialNumber :: Lens' CreateDevice (Maybe Text)

The serial number of the device.

Length Constraints: Maximum length of 128 characters.

createDevice_siteId :: Lens' CreateDevice (Maybe Text)

The ID of the site.

createDevice_description :: Lens' CreateDevice (Maybe Text)

A description of the device.

Length Constraints: Maximum length of 256 characters.

createDevice_tags :: Lens' CreateDevice (Maybe [Tag])

The tags to apply to the resource during creation.

createDevice_globalNetworkId :: Lens' CreateDevice Text

The ID of the global network.

createDeviceResponse_device :: Lens' CreateDeviceResponse (Maybe Device)

Information about the device.

createDeviceResponse_httpStatus :: Lens' CreateDeviceResponse Int

The response's http status code.

AssociateCustomerGateway

associateCustomerGateway_linkId :: Lens' AssociateCustomerGateway (Maybe Text)

The ID of the link.

associateCustomerGateway_customerGatewayArn :: Lens' AssociateCustomerGateway Text

The Amazon Resource Name (ARN) of the customer gateway. For more information, see Resources Defined by Amazon EC2.

associateCustomerGateway_globalNetworkId :: Lens' AssociateCustomerGateway Text

The ID of the global network.

associateCustomerGateway_deviceId :: Lens' AssociateCustomerGateway Text

The ID of the device.

associateCustomerGatewayResponse_customerGatewayAssociation :: Lens' AssociateCustomerGatewayResponse (Maybe CustomerGatewayAssociation)

The customer gateway association.

associateCustomerGatewayResponse_httpStatus :: Lens' AssociateCustomerGatewayResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the resource.

tagResource_tags :: Lens' TagResource [Tag]

The tags to apply to the specified resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

GetCustomerGatewayAssociations

getCustomerGatewayAssociations_nextToken :: Lens' GetCustomerGatewayAssociations (Maybe Text)

The token for the next page of results.

getCustomerGatewayAssociations_customerGatewayArns :: Lens' GetCustomerGatewayAssociations (Maybe [Text])

One or more customer gateway Amazon Resource Names (ARNs). For more information, see Resources Defined by Amazon EC2. The maximum is 10.

getCustomerGatewayAssociations_maxResults :: Lens' GetCustomerGatewayAssociations (Maybe Natural)

The maximum number of results to return.

getCustomerGatewayAssociations_globalNetworkId :: Lens' GetCustomerGatewayAssociations Text

The ID of the global network.

getCustomerGatewayAssociationsResponse_nextToken :: Lens' GetCustomerGatewayAssociationsResponse (Maybe Text)

The token for the next page of results.

getCustomerGatewayAssociationsResponse_customerGatewayAssociations :: Lens' GetCustomerGatewayAssociationsResponse (Maybe [CustomerGatewayAssociation])

The customer gateway associations.

getCustomerGatewayAssociationsResponse_httpStatus :: Lens' GetCustomerGatewayAssociationsResponse Int

The response's http status code.

GetTransitGatewayRegistrations

getTransitGatewayRegistrations_transitGatewayArns :: Lens' GetTransitGatewayRegistrations (Maybe [Text])

The Amazon Resource Names (ARNs) of one or more transit gateways. The maximum is 10.

getTransitGatewayRegistrations_nextToken :: Lens' GetTransitGatewayRegistrations (Maybe Text)

The token for the next page of results.

getTransitGatewayRegistrations_maxResults :: Lens' GetTransitGatewayRegistrations (Maybe Natural)

The maximum number of results to return.

getTransitGatewayRegistrations_globalNetworkId :: Lens' GetTransitGatewayRegistrations Text

The ID of the global network.

getTransitGatewayRegistrationsResponse_transitGatewayRegistrations :: Lens' GetTransitGatewayRegistrationsResponse (Maybe [TransitGatewayRegistration])

The transit gateway registrations.

getTransitGatewayRegistrationsResponse_nextToken :: Lens' GetTransitGatewayRegistrationsResponse (Maybe Text)

The token for the next page of results.

getTransitGatewayRegistrationsResponse_httpStatus :: Lens' GetTransitGatewayRegistrationsResponse Int

The response's http status code.

GetConnections

getConnections_connectionIds :: Lens' GetConnections (Maybe [Text])

One or more connection IDs.

getConnections_nextToken :: Lens' GetConnections (Maybe Text)

The token for the next page of results.

getConnections_deviceId :: Lens' GetConnections (Maybe Text)

The ID of the device.

getConnections_maxResults :: Lens' GetConnections (Maybe Natural)

The maximum number of results to return.

getConnections_globalNetworkId :: Lens' GetConnections Text

The ID of the global network.

getConnectionsResponse_connections :: Lens' GetConnectionsResponse (Maybe [Connection])

Information about the connections.

getConnectionsResponse_nextToken :: Lens' GetConnectionsResponse (Maybe Text)

The token to use for the next page of results.

getConnectionsResponse_httpStatus :: Lens' GetConnectionsResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the resource.

untagResource_tagKeys :: Lens' UntagResource [Text]

The tag keys to remove from the specified resource.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

GetSites

getSites_nextToken :: Lens' GetSites (Maybe Text)

The token for the next page of results.

getSites_siteIds :: Lens' GetSites (Maybe [Text])

One or more site IDs. The maximum is 10.

getSites_maxResults :: Lens' GetSites (Maybe Natural)

The maximum number of results to return.

getSites_globalNetworkId :: Lens' GetSites Text

The ID of the global network.

getSitesResponse_nextToken :: Lens' GetSitesResponse (Maybe Text)

The token for the next page of results.

getSitesResponse_sites :: Lens' GetSitesResponse (Maybe [Site])

The sites.

getSitesResponse_httpStatus :: Lens' GetSitesResponse Int

The response's http status code.

RegisterTransitGateway

registerTransitGateway_globalNetworkId :: Lens' RegisterTransitGateway Text

The ID of the global network.

registerTransitGateway_transitGatewayArn :: Lens' RegisterTransitGateway Text

The Amazon Resource Name (ARN) of the transit gateway. For more information, see Resources Defined by Amazon EC2.

registerTransitGatewayResponse_transitGatewayRegistration :: Lens' RegisterTransitGatewayResponse (Maybe TransitGatewayRegistration)

Information about the transit gateway registration.

registerTransitGatewayResponse_httpStatus :: Lens' RegisterTransitGatewayResponse Int

The response's http status code.

DeleteDevice

deleteDevice_globalNetworkId :: Lens' DeleteDevice Text

The ID of the global network.

deleteDevice_deviceId :: Lens' DeleteDevice Text

The ID of the device.

deleteDeviceResponse_device :: Lens' DeleteDeviceResponse (Maybe Device)

Information about the device.

deleteDeviceResponse_httpStatus :: Lens' DeleteDeviceResponse Int

The response's http status code.

UpdateDevice

updateDevice_vendor :: Lens' UpdateDevice (Maybe Text)

The vendor of the device.

Length Constraints: Maximum length of 128 characters.

updateDevice_location :: Lens' UpdateDevice (Maybe Location)

Undocumented member.

updateDevice_aWSLocation :: Lens' UpdateDevice (Maybe AWSLocation)

The AWS location of the device.

updateDevice_model :: Lens' UpdateDevice (Maybe Text)

The model of the device.

Length Constraints: Maximum length of 128 characters.

updateDevice_type :: Lens' UpdateDevice (Maybe Text)

The type of the device.

updateDevice_serialNumber :: Lens' UpdateDevice (Maybe Text)

The serial number of the device.

Length Constraints: Maximum length of 128 characters.

updateDevice_siteId :: Lens' UpdateDevice (Maybe Text)

The ID of the site.

updateDevice_description :: Lens' UpdateDevice (Maybe Text)

A description of the device.

Length Constraints: Maximum length of 256 characters.

updateDevice_globalNetworkId :: Lens' UpdateDevice Text

The ID of the global network.

updateDevice_deviceId :: Lens' UpdateDevice Text

The ID of the device.

updateDeviceResponse_device :: Lens' UpdateDeviceResponse (Maybe Device)

Information about the device.

updateDeviceResponse_httpStatus :: Lens' UpdateDeviceResponse Int

The response's http status code.

DeleteLink

deleteLink_globalNetworkId :: Lens' DeleteLink Text

The ID of the global network.

deleteLink_linkId :: Lens' DeleteLink Text

The ID of the link.

deleteLinkResponse_link :: Lens' DeleteLinkResponse (Maybe Link)

Information about the link.

deleteLinkResponse_httpStatus :: Lens' DeleteLinkResponse Int

The response's http status code.

UpdateLink

updateLink_bandwidth :: Lens' UpdateLink (Maybe Bandwidth)

The upload and download speed in Mbps.

updateLink_type :: Lens' UpdateLink (Maybe Text)

The type of the link.

Length Constraints: Maximum length of 128 characters.

updateLink_description :: Lens' UpdateLink (Maybe Text)

A description of the link.

Length Constraints: Maximum length of 256 characters.

updateLink_provider :: Lens' UpdateLink (Maybe Text)

The provider of the link.

Length Constraints: Maximum length of 128 characters.

updateLink_globalNetworkId :: Lens' UpdateLink Text

The ID of the global network.

updateLink_linkId :: Lens' UpdateLink Text

The ID of the link.

updateLinkResponse_link :: Lens' UpdateLinkResponse (Maybe Link)

Information about the link.

updateLinkResponse_httpStatus :: Lens' UpdateLinkResponse Int

The response's http status code.

Types

AWSLocation

aWSLocation_zone :: Lens' AWSLocation (Maybe Text)

The Zone the device is located in. This can be the ID of an Availability Zone, Local Zone, Wavelength Zone, or an Outpost.

aWSLocation_subnetArn :: Lens' AWSLocation (Maybe Text)

The Amazon Resource Name (ARN) of the subnet the device is located in.

Bandwidth

bandwidth_downloadSpeed :: Lens' Bandwidth (Maybe Int)

Download speed in Mbps.

bandwidth_uploadSpeed :: Lens' Bandwidth (Maybe Int)

Upload speed in Mbps.

Connection

connection_state :: Lens' Connection (Maybe ConnectionState)

The state of the connection.

connection_createdAt :: Lens' Connection (Maybe UTCTime)

The date and time that the connection was created.

connection_globalNetworkId :: Lens' Connection (Maybe Text)

The ID of the global network.

connection_connectionId :: Lens' Connection (Maybe Text)

The ID of the connection.

connection_connectedDeviceId :: Lens' Connection (Maybe Text)

The ID of the second device in the connection.

connection_connectedLinkId :: Lens' Connection (Maybe Text)

The ID of the link for the second device in the connection.

connection_linkId :: Lens' Connection (Maybe Text)

The ID of the link for the first device in the connection.

connection_deviceId :: Lens' Connection (Maybe Text)

The ID of the first device in the connection.

connection_connectionArn :: Lens' Connection (Maybe Text)

The Amazon Resource Name (ARN) of the connection.

connection_description :: Lens' Connection (Maybe Text)

The description of the connection.

connection_tags :: Lens' Connection (Maybe [Tag])

The tags for the connection.

CustomerGatewayAssociation

customerGatewayAssociation_state :: Lens' CustomerGatewayAssociation (Maybe CustomerGatewayAssociationState)

The association state.

customerGatewayAssociation_globalNetworkId :: Lens' CustomerGatewayAssociation (Maybe Text)

The ID of the global network.

customerGatewayAssociation_linkId :: Lens' CustomerGatewayAssociation (Maybe Text)

The ID of the link.

customerGatewayAssociation_deviceId :: Lens' CustomerGatewayAssociation (Maybe Text)

The ID of the device.

customerGatewayAssociation_customerGatewayArn :: Lens' CustomerGatewayAssociation (Maybe Text)

The Amazon Resource Name (ARN) of the customer gateway.

Device

device_vendor :: Lens' Device (Maybe Text)

The device vendor.

device_state :: Lens' Device (Maybe DeviceState)

The device state.

device_location :: Lens' Device (Maybe Location)

The site location.

device_createdAt :: Lens' Device (Maybe UTCTime)

The date and time that the site was created.

device_deviceArn :: Lens' Device (Maybe Text)

The Amazon Resource Name (ARN) of the device.

device_aWSLocation :: Lens' Device (Maybe AWSLocation)

The AWS location of the device.

device_globalNetworkId :: Lens' Device (Maybe Text)

The ID of the global network.

device_model :: Lens' Device (Maybe Text)

The device model.

device_deviceId :: Lens' Device (Maybe Text)

The ID of the device.

device_type :: Lens' Device (Maybe Text)

The device type.

device_serialNumber :: Lens' Device (Maybe Text)

The device serial number.

device_siteId :: Lens' Device (Maybe Text)

The site ID.

device_description :: Lens' Device (Maybe Text)

The description of the device.

device_tags :: Lens' Device (Maybe [Tag])

The tags for the device.

GlobalNetwork

globalNetwork_state :: Lens' GlobalNetwork (Maybe GlobalNetworkState)

The state of the global network.

globalNetwork_createdAt :: Lens' GlobalNetwork (Maybe UTCTime)

The date and time that the global network was created.

globalNetwork_globalNetworkArn :: Lens' GlobalNetwork (Maybe Text)

The Amazon Resource Name (ARN) of the global network.

globalNetwork_globalNetworkId :: Lens' GlobalNetwork (Maybe Text)

The ID of the global network.

globalNetwork_description :: Lens' GlobalNetwork (Maybe Text)

The description of the global network.

globalNetwork_tags :: Lens' GlobalNetwork (Maybe [Tag])

The tags for the global network.

Link

link_state :: Lens' Link (Maybe LinkState)

The state of the link.

link_linkArn :: Lens' Link (Maybe Text)

The Amazon Resource Name (ARN) of the link.

link_createdAt :: Lens' Link (Maybe UTCTime)

The date and time that the link was created.

link_globalNetworkId :: Lens' Link (Maybe Text)

The ID of the global network.

link_bandwidth :: Lens' Link (Maybe Bandwidth)

The bandwidth for the link.

link_linkId :: Lens' Link (Maybe Text)

The ID of the link.

link_type :: Lens' Link (Maybe Text)

The type of the link.

link_siteId :: Lens' Link (Maybe Text)

The ID of the site.

link_description :: Lens' Link (Maybe Text)

The description of the link.

link_provider :: Lens' Link (Maybe Text)

The provider of the link.

link_tags :: Lens' Link (Maybe [Tag])

The tags for the link.

LinkAssociation

linkAssociation_globalNetworkId :: Lens' LinkAssociation (Maybe Text)

The ID of the global network.

linkAssociation_linkId :: Lens' LinkAssociation (Maybe Text)

The ID of the link.

linkAssociation_deviceId :: Lens' LinkAssociation (Maybe Text)

The device ID for the link association.

linkAssociation_linkAssociationState :: Lens' LinkAssociation (Maybe LinkAssociationState)

The state of the association.

Location

location_latitude :: Lens' Location (Maybe Text)

The latitude.

location_address :: Lens' Location (Maybe Text)

The physical address.

location_longitude :: Lens' Location (Maybe Text)

The longitude.

Site

site_state :: Lens' Site (Maybe SiteState)

The state of the site.

site_location :: Lens' Site (Maybe Location)

The location of the site.

site_createdAt :: Lens' Site (Maybe UTCTime)

The date and time that the site was created.

site_globalNetworkId :: Lens' Site (Maybe Text)

The ID of the global network.

site_siteId :: Lens' Site (Maybe Text)

The ID of the site.

site_siteArn :: Lens' Site (Maybe Text)

The Amazon Resource Name (ARN) of the site.

site_description :: Lens' Site (Maybe Text)

The description of the site.

site_tags :: Lens' Site (Maybe [Tag])

The tags for the site.

Tag

tag_value :: Lens' Tag (Maybe Text)

The tag value.

Length Constraints: Maximum length of 256 characters.

tag_key :: Lens' Tag (Maybe Text)

The tag key.

Length Constraints: Maximum length of 128 characters.

TransitGatewayConnectPeerAssociation

transitGatewayConnectPeerAssociation_state :: Lens' TransitGatewayConnectPeerAssociation (Maybe TransitGatewayConnectPeerAssociationState)

The state of the association.

transitGatewayConnectPeerAssociation_globalNetworkId :: Lens' TransitGatewayConnectPeerAssociation (Maybe Text)

The ID of the global network.

transitGatewayConnectPeerAssociation_transitGatewayConnectPeerArn :: Lens' TransitGatewayConnectPeerAssociation (Maybe Text)

The Amazon Resource Name (ARN) of the transit gateway Connect peer.

transitGatewayConnectPeerAssociation_linkId :: Lens' TransitGatewayConnectPeerAssociation (Maybe Text)

The ID of the link.

transitGatewayConnectPeerAssociation_deviceId :: Lens' TransitGatewayConnectPeerAssociation (Maybe Text)

The ID of the device.

TransitGatewayRegistration

transitGatewayRegistration_state :: Lens' TransitGatewayRegistration (Maybe TransitGatewayRegistrationStateReason)

The state of the transit gateway registration.

transitGatewayRegistration_globalNetworkId :: Lens' TransitGatewayRegistration (Maybe Text)

The ID of the global network.

transitGatewayRegistration_transitGatewayArn :: Lens' TransitGatewayRegistration (Maybe Text)

The Amazon Resource Name (ARN) of the transit gateway.

TransitGatewayRegistrationStateReason

transitGatewayRegistrationStateReason_code :: Lens' TransitGatewayRegistrationStateReason (Maybe TransitGatewayRegistrationState)

The code for the state reason.

transitGatewayRegistrationStateReason_message :: Lens' TransitGatewayRegistrationStateReason (Maybe Text)

The message for the state reason.