Network.AWS.NetworkManager.Types.AWSLocation

Description

Documentation

data AWSLocation

Specifies a location in AWS.

See: newAWSLocation smart constructor.

Constructors

AWSLocation'

Fields

zone :: Maybe Text

The Zone the device is located in. This can be the ID of an Availability Zone, Local Zone, Wavelength Zone, or an Outpost.

subnetArn :: Maybe Text

The Amazon Resource Name (ARN) of the subnet the device is located in.

newAWSLocation :: AWSLocation

Create a value of AWSLocation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:zone:AWSLocation', aWSLocation_zone - The Zone the device is located in. This can be the ID of an Availability Zone, Local Zone, Wavelength Zone, or an Outpost.

$sel:subnetArn:AWSLocation', aWSLocation_subnetArn - The Amazon Resource Name (ARN) of the subnet the device is located in.

aWSLocation_zone :: Lens' AWSLocation (Maybe Text)

The Zone the device is located in. This can be the ID of an Availability Zone, Local Zone, Wavelength Zone, or an Outpost.

aWSLocation_subnetArn :: Lens' AWSLocation (Maybe Text)

The Amazon Resource Name (ARN) of the subnet the device is located in.