Network.AWS.NetworkManager.Types.Connection

Description

Documentation

data Connection

Describes a connection.

See: newConnection smart constructor.

Constructors

Connection'

Fields

state :: Maybe ConnectionState

The state of the connection.

createdAt :: Maybe POSIX

The date and time that the connection was created.

globalNetworkId :: Maybe Text

The ID of the global network.

connectionId :: Maybe Text

The ID of the connection.

connectedDeviceId :: Maybe Text

The ID of the second device in the connection.

connectedLinkId :: Maybe Text

The ID of the link for the second device in the connection.

linkId :: Maybe Text

The ID of the link for the first device in the connection.

deviceId :: Maybe Text

The ID of the first device in the connection.

connectionArn :: Maybe Text

The Amazon Resource Name (ARN) of the connection.

description :: Maybe Text

The description of the connection.

tags :: Maybe [Tag]

The tags for the connection.

newConnection :: Connection

Create a value of Connection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:Connection', connection_state - The state of the connection.

$sel:createdAt:Connection', connection_createdAt - The date and time that the connection was created.

$sel:globalNetworkId:Connection', connection_globalNetworkId - The ID of the global network.

$sel:connectionId:Connection', connection_connectionId - The ID of the connection.

$sel:connectedDeviceId:Connection', connection_connectedDeviceId - The ID of the second device in the connection.

$sel:connectedLinkId:Connection', connection_connectedLinkId - The ID of the link for the second device in the connection.

$sel:linkId:Connection', connection_linkId - The ID of the link for the first device in the connection.

$sel:deviceId:Connection', connection_deviceId - The ID of the first device in the connection.

$sel:connectionArn:Connection', connection_connectionArn - The Amazon Resource Name (ARN) of the connection.

$sel:description:Connection', connection_description - The description of the connection.

$sel:tags:Connection', connection_tags - The tags for the connection.

connection_state :: Lens' Connection (Maybe ConnectionState)

The state of the connection.

connection_createdAt :: Lens' Connection (Maybe UTCTime)

The date and time that the connection was created.

connection_globalNetworkId :: Lens' Connection (Maybe Text)

The ID of the global network.

connection_connectionId :: Lens' Connection (Maybe Text)

The ID of the connection.

connection_connectedDeviceId :: Lens' Connection (Maybe Text)

The ID of the second device in the connection.

connection_connectedLinkId :: Lens' Connection (Maybe Text)

The ID of the link for the second device in the connection.

connection_linkId :: Lens' Connection (Maybe Text)

The ID of the link for the first device in the connection.

connection_deviceId :: Lens' Connection (Maybe Text)

The ID of the first device in the connection.

connection_connectionArn :: Lens' Connection (Maybe Text)

The Amazon Resource Name (ARN) of the connection.

connection_description :: Lens' Connection (Maybe Text)

The description of the connection.

connection_tags :: Lens' Connection (Maybe [Tag])

The tags for the connection.