Network.AWS.NetworkManager.Types.CustomerGatewayAssociation

Description

Documentation

data CustomerGatewayAssociation

Describes the association between a customer gateway, a device, and a link.

See: newCustomerGatewayAssociation smart constructor.

Constructors

CustomerGatewayAssociation'

Fields

state :: Maybe CustomerGatewayAssociationState

The association state.

globalNetworkId :: Maybe Text

The ID of the global network.

linkId :: Maybe Text

The ID of the link.

deviceId :: Maybe Text

The ID of the device.

customerGatewayArn :: Maybe Text

The Amazon Resource Name (ARN) of the customer gateway.

newCustomerGatewayAssociation :: CustomerGatewayAssociation

Create a value of CustomerGatewayAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:CustomerGatewayAssociation', customerGatewayAssociation_state - The association state.

$sel:globalNetworkId:CustomerGatewayAssociation', customerGatewayAssociation_globalNetworkId - The ID of the global network.

$sel:linkId:CustomerGatewayAssociation', customerGatewayAssociation_linkId - The ID of the link.

$sel:deviceId:CustomerGatewayAssociation', customerGatewayAssociation_deviceId - The ID of the device.

$sel:customerGatewayArn:CustomerGatewayAssociation', customerGatewayAssociation_customerGatewayArn - The Amazon Resource Name (ARN) of the customer gateway.

customerGatewayAssociation_state :: Lens' CustomerGatewayAssociation (Maybe CustomerGatewayAssociationState)

The association state.

customerGatewayAssociation_globalNetworkId :: Lens' CustomerGatewayAssociation (Maybe Text)

The ID of the global network.

customerGatewayAssociation_linkId :: Lens' CustomerGatewayAssociation (Maybe Text)

The ID of the link.

customerGatewayAssociation_deviceId :: Lens' CustomerGatewayAssociation (Maybe Text)

The ID of the device.

customerGatewayAssociation_customerGatewayArn :: Lens' CustomerGatewayAssociation (Maybe Text)

The Amazon Resource Name (ARN) of the customer gateway.