Network.AWS.NetworkManager.Types.Device

Description

Documentation

data Device

Describes a device.

See: newDevice smart constructor.

Constructors

Device'

Fields

vendor :: Maybe Text

The device vendor.

state :: Maybe DeviceState

The device state.

location :: Maybe (Sensitive Location)

The site location.

createdAt :: Maybe POSIX

The date and time that the site was created.

deviceArn :: Maybe Text

The Amazon Resource Name (ARN) of the device.

aWSLocation :: Maybe AWSLocation

The AWS location of the device.

globalNetworkId :: Maybe Text

The ID of the global network.

model :: Maybe Text

The device model.

deviceId :: Maybe Text

The ID of the device.

type' :: Maybe Text

The device type.

serialNumber :: Maybe Text

The device serial number.

siteId :: Maybe Text

The site ID.

description :: Maybe Text

The description of the device.

tags :: Maybe [Tag]

The tags for the device.

newDevice :: Device

Create a value of Device with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vendor:Device', device_vendor - The device vendor.

$sel:state:Device', device_state - The device state.

$sel:location:Device', device_location - The site location.

$sel:createdAt:Device', device_createdAt - The date and time that the site was created.

$sel:deviceArn:Device', device_deviceArn - The Amazon Resource Name (ARN) of the device.

$sel:aWSLocation:Device', device_aWSLocation - The AWS location of the device.

$sel:globalNetworkId:Device', device_globalNetworkId - The ID of the global network.

$sel:model:Device', device_model - The device model.

$sel:deviceId:Device', device_deviceId - The ID of the device.

$sel:type':Device', device_type - The device type.

$sel:serialNumber:Device', device_serialNumber - The device serial number.

$sel:siteId:Device', device_siteId - The site ID.

$sel:description:Device', device_description - The description of the device.

$sel:tags:Device', device_tags - The tags for the device.

device_vendor :: Lens' Device (Maybe Text)

The device vendor.

device_state :: Lens' Device (Maybe DeviceState)

The device state.

device_location :: Lens' Device (Maybe Location)

The site location.

device_createdAt :: Lens' Device (Maybe UTCTime)

The date and time that the site was created.

device_deviceArn :: Lens' Device (Maybe Text)

The Amazon Resource Name (ARN) of the device.

device_aWSLocation :: Lens' Device (Maybe AWSLocation)

The AWS location of the device.

device_globalNetworkId :: Lens' Device (Maybe Text)

The ID of the global network.

device_model :: Lens' Device (Maybe Text)

The device model.

device_deviceId :: Lens' Device (Maybe Text)

The ID of the device.

device_type :: Lens' Device (Maybe Text)

The device type.

device_serialNumber :: Lens' Device (Maybe Text)

The device serial number.

device_siteId :: Lens' Device (Maybe Text)

The site ID.

device_description :: Lens' Device (Maybe Text)

The description of the device.

device_tags :: Lens' Device (Maybe [Tag])

The tags for the device.