Network.AWS.NetworkManager.Types.Link

Description

Documentation

data Link

Describes a link.

See: newLink smart constructor.

Constructors

Link'

Fields

state :: Maybe LinkState

The state of the link.

linkArn :: Maybe Text

The Amazon Resource Name (ARN) of the link.

createdAt :: Maybe POSIX

The date and time that the link was created.

globalNetworkId :: Maybe Text

The ID of the global network.

bandwidth :: Maybe Bandwidth

The bandwidth for the link.

linkId :: Maybe Text

The ID of the link.

type' :: Maybe Text

The type of the link.

siteId :: Maybe Text

The ID of the site.

description :: Maybe Text

The description of the link.

provider :: Maybe Text

The provider of the link.

tags :: Maybe [Tag]

The tags for the link.

newLink :: Link

Create a value of Link with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:Link', link_state - The state of the link.

$sel:linkArn:Link', link_linkArn - The Amazon Resource Name (ARN) of the link.

$sel:createdAt:Link', link_createdAt - The date and time that the link was created.

$sel:globalNetworkId:Link', link_globalNetworkId - The ID of the global network.

$sel:bandwidth:Link', link_bandwidth - The bandwidth for the link.

$sel:linkId:Link', link_linkId - The ID of the link.

$sel:type':Link', link_type - The type of the link.

$sel:siteId:Link', link_siteId - The ID of the site.

$sel:description:Link', link_description - The description of the link.

$sel:provider:Link', link_provider - The provider of the link.

$sel:tags:Link', link_tags - The tags for the link.

link_state :: Lens' Link (Maybe LinkState)

The state of the link.

link_linkArn :: Lens' Link (Maybe Text)

The Amazon Resource Name (ARN) of the link.

link_createdAt :: Lens' Link (Maybe UTCTime)

The date and time that the link was created.

link_globalNetworkId :: Lens' Link (Maybe Text)

The ID of the global network.

link_bandwidth :: Lens' Link (Maybe Bandwidth)

The bandwidth for the link.

link_linkId :: Lens' Link (Maybe Text)

The ID of the link.

link_type :: Lens' Link (Maybe Text)

The type of the link.

link_siteId :: Lens' Link (Maybe Text)

The ID of the site.

link_description :: Lens' Link (Maybe Text)

The description of the link.

link_provider :: Lens' Link (Maybe Text)

The provider of the link.

link_tags :: Lens' Link (Maybe [Tag])

The tags for the link.