Network.AWS.Nimble.CreateStreamingSession

Description

Creates a streaming session in a studio.

After invoking this operation, you must poll GetStreamingSession until the streaming session is in state READY.

Creating a Request

data CreateStreamingSession

A collection of streaming sessions.

See: newCreateStreamingSession smart constructor.

Constructors

CreateStreamingSession'

Fields

ownedBy :: Maybe Text

The user ID of the user that owns the streaming session.

clientToken :: Maybe Text

To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

ec2InstanceType :: Maybe StreamingInstanceType

The EC2 Instance type used for the streaming session.

launchProfileId :: Maybe Text

The launch profile ID.

streamingImageId :: Maybe Text

The ID of the streaming image.

tags :: Maybe (HashMap Text Text)

A collection of labels, in the form of key:value pairs, that apply to this resource.

studioId :: Text

The studio ID.

newCreateStreamingSession

Arguments

:: Text
$sel:studioId:CreateStreamingSession'

-> CreateStreamingSession

Create a value of CreateStreamingSession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownedBy:CreateStreamingSession', createStreamingSession_ownedBy - The user ID of the user that owns the streaming session.

$sel:clientToken:CreateStreamingSession', createStreamingSession_clientToken - To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

$sel:ec2InstanceType:CreateStreamingSession', createStreamingSession_ec2InstanceType - The EC2 Instance type used for the streaming session.

$sel:launchProfileId:CreateStreamingSession', createStreamingSession_launchProfileId - The launch profile ID.

$sel:streamingImageId:CreateStreamingSession', createStreamingSession_streamingImageId - The ID of the streaming image.

$sel:tags:CreateStreamingSession', createStreamingSession_tags - A collection of labels, in the form of key:value pairs, that apply to this resource.

$sel:studioId:CreateStreamingSession', createStreamingSession_studioId - The studio ID.

Request Lenses

createStreamingSession_ownedBy :: Lens' CreateStreamingSession (Maybe Text)

The user ID of the user that owns the streaming session.

createStreamingSession_clientToken :: Lens' CreateStreamingSession (Maybe Text)

To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

createStreamingSession_ec2InstanceType :: Lens' CreateStreamingSession (Maybe StreamingInstanceType)

The EC2 Instance type used for the streaming session.

createStreamingSession_launchProfileId :: Lens' CreateStreamingSession (Maybe Text)

The launch profile ID.

createStreamingSession_streamingImageId :: Lens' CreateStreamingSession (Maybe Text)

The ID of the streaming image.

createStreamingSession_tags :: Lens' CreateStreamingSession (Maybe (HashMap Text Text))

A collection of labels, in the form of key:value pairs, that apply to this resource.

createStreamingSession_studioId :: Lens' CreateStreamingSession Text

The studio ID.

Destructuring the Response

data CreateStreamingSessionResponse

See: newCreateStreamingSessionResponse smart constructor.

Constructors

CreateStreamingSessionResponse'

Fields

session :: Maybe StreamingSession

The session.

httpStatus :: Int

The response's http status code.

newCreateStreamingSessionResponse

Arguments

:: Int
$sel:httpStatus:CreateStreamingSessionResponse'

-> CreateStreamingSessionResponse

Create a value of CreateStreamingSessionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:session:CreateStreamingSessionResponse', createStreamingSessionResponse_session - The session.

$sel:httpStatus:CreateStreamingSessionResponse', createStreamingSessionResponse_httpStatus - The response's http status code.

Response Lenses

createStreamingSessionResponse_session :: Lens' CreateStreamingSessionResponse (Maybe StreamingSession)

The session.

createStreamingSessionResponse_httpStatus :: Lens' CreateStreamingSessionResponse Int

The response's http status code.