Network.AWS.Nimble.DeleteLaunchProfileMember

Description

Delete a user from launch profile membership.

Creating a Request

data DeleteLaunchProfileMember

See: newDeleteLaunchProfileMember smart constructor.

Constructors

DeleteLaunchProfileMember'

Fields

clientToken :: Maybe Text

To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

studioId :: Text

The studio ID.

principalId :: Text

The principal ID.

launchProfileId :: Text

The launch profile ID.

newDeleteLaunchProfileMember

Arguments

:: Text
$sel:studioId:DeleteLaunchProfileMember'

-> Text
$sel:principalId:DeleteLaunchProfileMember'

-> Text
$sel:launchProfileId:DeleteLaunchProfileMember'

-> DeleteLaunchProfileMember

Create a value of DeleteLaunchProfileMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:DeleteLaunchProfileMember', deleteLaunchProfileMember_clientToken - To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

$sel:studioId:DeleteLaunchProfileMember', deleteLaunchProfileMember_studioId - The studio ID.

$sel:principalId:DeleteLaunchProfileMember', deleteLaunchProfileMember_principalId - The principal ID.

$sel:launchProfileId:DeleteLaunchProfileMember', deleteLaunchProfileMember_launchProfileId - The launch profile ID.

Request Lenses

deleteLaunchProfileMember_clientToken :: Lens' DeleteLaunchProfileMember (Maybe Text)

To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

deleteLaunchProfileMember_studioId :: Lens' DeleteLaunchProfileMember Text

The studio ID.

deleteLaunchProfileMember_principalId :: Lens' DeleteLaunchProfileMember Text

The principal ID.

deleteLaunchProfileMember_launchProfileId :: Lens' DeleteLaunchProfileMember Text

The launch profile ID.

Destructuring the Response

data DeleteLaunchProfileMemberResponse

See: newDeleteLaunchProfileMemberResponse smart constructor.

Constructors

DeleteLaunchProfileMemberResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteLaunchProfileMemberResponse

Arguments

:: Int
$sel:httpStatus:DeleteLaunchProfileMemberResponse'

-> DeleteLaunchProfileMemberResponse

Create a value of DeleteLaunchProfileMemberResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteLaunchProfileMemberResponse', deleteLaunchProfileMemberResponse_httpStatus - The response's http status code.

Response Lenses

deleteLaunchProfileMemberResponse_httpStatus :: Lens' DeleteLaunchProfileMemberResponse Int

The response's http status code.