Network.AWS.Nimble.DeleteStreamingImage

Description

Delete streaming image.

Creating a Request

data DeleteStreamingImage

See: newDeleteStreamingImage smart constructor.

Constructors

DeleteStreamingImage'

Fields

clientToken :: Maybe Text

To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

studioId :: Text

The studio ID.

streamingImageId :: Text

The streaming image ID.

newDeleteStreamingImage

Arguments

:: Text
$sel:studioId:DeleteStreamingImage'

-> Text
$sel:streamingImageId:DeleteStreamingImage'

-> DeleteStreamingImage

Create a value of DeleteStreamingImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:DeleteStreamingImage', deleteStreamingImage_clientToken - To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

$sel:studioId:DeleteStreamingImage', deleteStreamingImage_studioId - The studio ID.

$sel:streamingImageId:DeleteStreamingImage', deleteStreamingImage_streamingImageId - The streaming image ID.

Request Lenses

deleteStreamingImage_clientToken :: Lens' DeleteStreamingImage (Maybe Text)

To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

deleteStreamingImage_studioId :: Lens' DeleteStreamingImage Text

The studio ID.

deleteStreamingImage_streamingImageId :: Lens' DeleteStreamingImage Text

The streaming image ID.

Destructuring the Response

data DeleteStreamingImageResponse

See: newDeleteStreamingImageResponse smart constructor.

Constructors

DeleteStreamingImageResponse'

Fields

streamingImage :: Maybe StreamingImage

The streaming image.

httpStatus :: Int

The response's http status code.

newDeleteStreamingImageResponse

Arguments

:: Int
$sel:httpStatus:DeleteStreamingImageResponse'

-> DeleteStreamingImageResponse

Create a value of DeleteStreamingImageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamingImage:DeleteStreamingImageResponse', deleteStreamingImageResponse_streamingImage - The streaming image.

$sel:httpStatus:DeleteStreamingImageResponse', deleteStreamingImageResponse_httpStatus - The response's http status code.

Response Lenses

deleteStreamingImageResponse_streamingImage :: Lens' DeleteStreamingImageResponse (Maybe StreamingImage)

The streaming image.

deleteStreamingImageResponse_httpStatus :: Lens' DeleteStreamingImageResponse Int

The response's http status code.