Network.AWS.Nimble.GetLaunchProfileMember

Description

Get a user persona in launch profile membership.

Creating a Request

data GetLaunchProfileMember

See: newGetLaunchProfileMember smart constructor.

Constructors

GetLaunchProfileMember'

Fields

studioId :: Text

The studio ID.

principalId :: Text

The principal ID.

launchProfileId :: Text

The launch profile ID.

newGetLaunchProfileMember

Arguments

:: Text
$sel:studioId:GetLaunchProfileMember'

-> Text
$sel:principalId:GetLaunchProfileMember'

-> Text
$sel:launchProfileId:GetLaunchProfileMember'

-> GetLaunchProfileMember

Create a value of GetLaunchProfileMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:studioId:GetLaunchProfileMember', getLaunchProfileMember_studioId - The studio ID.

$sel:principalId:GetLaunchProfileMember', getLaunchProfileMember_principalId - The principal ID.

$sel:launchProfileId:GetLaunchProfileMember', getLaunchProfileMember_launchProfileId - The launch profile ID.

Request Lenses

getLaunchProfileMember_studioId :: Lens' GetLaunchProfileMember Text

The studio ID.

getLaunchProfileMember_principalId :: Lens' GetLaunchProfileMember Text

The principal ID.

getLaunchProfileMember_launchProfileId :: Lens' GetLaunchProfileMember Text

The launch profile ID.

Destructuring the Response

data GetLaunchProfileMemberResponse

See: newGetLaunchProfileMemberResponse smart constructor.

Constructors

GetLaunchProfileMemberResponse'

Fields

member :: Maybe LaunchProfileMembership

The member.

httpStatus :: Int

The response's http status code.

newGetLaunchProfileMemberResponse

Arguments

:: Int
$sel:httpStatus:GetLaunchProfileMemberResponse'

-> GetLaunchProfileMemberResponse

Create a value of GetLaunchProfileMemberResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:member:GetLaunchProfileMemberResponse', getLaunchProfileMemberResponse_member - The member.

$sel:httpStatus:GetLaunchProfileMemberResponse', getLaunchProfileMemberResponse_httpStatus - The response's http status code.

Response Lenses

getLaunchProfileMemberResponse_member :: Lens' GetLaunchProfileMemberResponse (Maybe LaunchProfileMembership)

The member.

getLaunchProfileMemberResponse_httpStatus :: Lens' GetLaunchProfileMemberResponse Int

The response's http status code.