Network.AWS.Nimble.ListEulas

Description

List Eulas.

This operation returns paginated results.

Creating a Request

data ListEulas

See: newListEulas smart constructor.

Constructors

ListEulas'

Fields

nextToken :: Maybe Text

The token for the next set of results, or null if there are no more results.

eulaIds :: Maybe [Text]

A collection of EULA IDs.

newListEulas :: ListEulas

Create a value of ListEulas with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEulas', listEulas_nextToken - The token for the next set of results, or null if there are no more results.

$sel:eulaIds:ListEulas', listEulas_eulaIds - A collection of EULA IDs.

Request Lenses

listEulas_nextToken :: Lens' ListEulas (Maybe Text)

The token for the next set of results, or null if there are no more results.

listEulas_eulaIds :: Lens' ListEulas (Maybe [Text])

A collection of EULA IDs.

Destructuring the Response

data ListEulasResponse

See: newListEulasResponse smart constructor.

Constructors

ListEulasResponse'

Fields

eulas :: Maybe [Eula]

A collection of EULA resources.

nextToken :: Maybe Text

The token for the next set of results, or null if there are no more results.

httpStatus :: Int

The response's http status code.

newListEulasResponse

Arguments

:: Int
$sel:httpStatus:ListEulasResponse'

-> ListEulasResponse

Create a value of ListEulasResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eulas:ListEulasResponse', listEulasResponse_eulas - A collection of EULA resources.

$sel:nextToken:ListEulas', listEulasResponse_nextToken - The token for the next set of results, or null if there are no more results.

$sel:httpStatus:ListEulasResponse', listEulasResponse_httpStatus - The response's http status code.

Response Lenses

listEulasResponse_eulas :: Lens' ListEulasResponse (Maybe [Eula])

A collection of EULA resources.

listEulasResponse_nextToken :: Lens' ListEulasResponse (Maybe Text)

The token for the next set of results, or null if there are no more results.

listEulasResponse_httpStatus :: Lens' ListEulasResponse Int

The response's http status code.