Network.AWS.Nimble.ListLaunchProfiles

Description

List all the launch profiles a studio.

This operation returns paginated results.

Creating a Request

data ListLaunchProfiles

See: newListLaunchProfiles smart constructor.

Constructors

ListLaunchProfiles'

Fields

states :: Maybe [Text]

A list of states.

principalId :: Maybe Text

The principal ID.

nextToken :: Maybe Text

The token for the next set of results, or null if there are no more results.

maxResults :: Maybe Natural

The maximum number of results to be returned per request.

studioId :: Text

The studio ID.

newListLaunchProfiles

Arguments

:: Text
$sel:studioId:ListLaunchProfiles'

-> ListLaunchProfiles

Create a value of ListLaunchProfiles with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:states:ListLaunchProfiles', listLaunchProfiles_states - A list of states.

$sel:principalId:ListLaunchProfiles', listLaunchProfiles_principalId - The principal ID.

$sel:nextToken:ListLaunchProfiles', listLaunchProfiles_nextToken - The token for the next set of results, or null if there are no more results.

$sel:maxResults:ListLaunchProfiles', listLaunchProfiles_maxResults - The maximum number of results to be returned per request.

$sel:studioId:ListLaunchProfiles', listLaunchProfiles_studioId - The studio ID.

Request Lenses

listLaunchProfiles_states :: Lens' ListLaunchProfiles (Maybe [Text])

A list of states.

listLaunchProfiles_principalId :: Lens' ListLaunchProfiles (Maybe Text)

The principal ID.

listLaunchProfiles_nextToken :: Lens' ListLaunchProfiles (Maybe Text)

The token for the next set of results, or null if there are no more results.

listLaunchProfiles_maxResults :: Lens' ListLaunchProfiles (Maybe Natural)

The maximum number of results to be returned per request.

listLaunchProfiles_studioId :: Lens' ListLaunchProfiles Text

The studio ID.

Destructuring the Response

data ListLaunchProfilesResponse

See: newListLaunchProfilesResponse smart constructor.

Constructors

ListLaunchProfilesResponse'

Fields

launchProfiles :: Maybe [LaunchProfile]

A collection of launch profiles.

nextToken :: Maybe Text

The token for the next set of results, or null if there are no more results.

httpStatus :: Int

The response's http status code.

newListLaunchProfilesResponse

Arguments

:: Int
$sel:httpStatus:ListLaunchProfilesResponse'

-> ListLaunchProfilesResponse

Create a value of ListLaunchProfilesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:launchProfiles:ListLaunchProfilesResponse', listLaunchProfilesResponse_launchProfiles - A collection of launch profiles.

$sel:nextToken:ListLaunchProfiles', listLaunchProfilesResponse_nextToken - The token for the next set of results, or null if there are no more results.

$sel:httpStatus:ListLaunchProfilesResponse', listLaunchProfilesResponse_httpStatus - The response's http status code.

Response Lenses

listLaunchProfilesResponse_launchProfiles :: Lens' ListLaunchProfilesResponse (Maybe [LaunchProfile])

A collection of launch profiles.

listLaunchProfilesResponse_nextToken :: Lens' ListLaunchProfilesResponse (Maybe Text)

The token for the next set of results, or null if there are no more results.

listLaunchProfilesResponse_httpStatus :: Lens' ListLaunchProfilesResponse Int

The response's http status code.