Network.AWS.Nimble.ListStreamingSessions

Description

Lists the streaming image resources in a studio.

This operation returns paginated results.

Creating a Request

data ListStreamingSessions

See: newListStreamingSessions smart constructor.

Constructors

ListStreamingSessions'

Fields

ownedBy :: Maybe Text

The user ID of the user that owns the streaming session.

createdBy :: Maybe Text

The user ID of the user that created the streaming session.

nextToken :: Maybe Text

The token for the next set of results, or null if there are no more results.

sessionIds :: Maybe Text

A collection of session IDs.

studioId :: Text

The studio ID.

newListStreamingSessions

Arguments

:: Text
$sel:studioId:ListStreamingSessions'

-> ListStreamingSessions

Create a value of ListStreamingSessions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownedBy:ListStreamingSessions', listStreamingSessions_ownedBy - The user ID of the user that owns the streaming session.

$sel:createdBy:ListStreamingSessions', listStreamingSessions_createdBy - The user ID of the user that created the streaming session.

$sel:nextToken:ListStreamingSessions', listStreamingSessions_nextToken - The token for the next set of results, or null if there are no more results.

$sel:sessionIds:ListStreamingSessions', listStreamingSessions_sessionIds - A collection of session IDs.

$sel:studioId:ListStreamingSessions', listStreamingSessions_studioId - The studio ID.

Request Lenses

listStreamingSessions_ownedBy :: Lens' ListStreamingSessions (Maybe Text)

The user ID of the user that owns the streaming session.

listStreamingSessions_createdBy :: Lens' ListStreamingSessions (Maybe Text)

The user ID of the user that created the streaming session.

listStreamingSessions_nextToken :: Lens' ListStreamingSessions (Maybe Text)

The token for the next set of results, or null if there are no more results.

listStreamingSessions_sessionIds :: Lens' ListStreamingSessions (Maybe Text)

A collection of session IDs.

listStreamingSessions_studioId :: Lens' ListStreamingSessions Text

The studio ID.

Destructuring the Response

data ListStreamingSessionsResponse

See: newListStreamingSessionsResponse smart constructor.

Constructors

ListStreamingSessionsResponse'

Fields

nextToken :: Maybe Text

The token for the next set of results, or null if there are no more results.

sessions :: Maybe [StreamingSession]

A collection of streaming sessions.

httpStatus :: Int

The response's http status code.

newListStreamingSessionsResponse

Arguments

:: Int
$sel:httpStatus:ListStreamingSessionsResponse'

-> ListStreamingSessionsResponse

Create a value of ListStreamingSessionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListStreamingSessions', listStreamingSessionsResponse_nextToken - The token for the next set of results, or null if there are no more results.

$sel:sessions:ListStreamingSessionsResponse', listStreamingSessionsResponse_sessions - A collection of streaming sessions.

$sel:httpStatus:ListStreamingSessionsResponse', listStreamingSessionsResponse_httpStatus - The response's http status code.

Response Lenses

listStreamingSessionsResponse_nextToken :: Lens' ListStreamingSessionsResponse (Maybe Text)

The token for the next set of results, or null if there are no more results.

listStreamingSessionsResponse_sessions :: Lens' ListStreamingSessionsResponse (Maybe [StreamingSession])

A collection of streaming sessions.

listStreamingSessionsResponse_httpStatus :: Lens' ListStreamingSessionsResponse Int

The response's http status code.