Network.AWS.Nimble.ListStudioMembers

Description

Get all users in a given studio membership.

This operation returns paginated results.

Creating a Request

data ListStudioMembers

See: newListStudioMembers smart constructor.

Constructors

ListStudioMembers'

Fields

nextToken :: Maybe Text

The token for the next set of results, or null if there are no more results.

maxResults :: Maybe Natural

The maximum number of results to be returned per request.

studioId :: Text

The studio ID.

newListStudioMembers

Arguments

:: Text
$sel:studioId:ListStudioMembers'

-> ListStudioMembers

Create a value of ListStudioMembers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListStudioMembers', listStudioMembers_nextToken - The token for the next set of results, or null if there are no more results.

$sel:maxResults:ListStudioMembers', listStudioMembers_maxResults - The maximum number of results to be returned per request.

$sel:studioId:ListStudioMembers', listStudioMembers_studioId - The studio ID.

Request Lenses

listStudioMembers_nextToken :: Lens' ListStudioMembers (Maybe Text)

The token for the next set of results, or null if there are no more results.

listStudioMembers_maxResults :: Lens' ListStudioMembers (Maybe Natural)

The maximum number of results to be returned per request.

listStudioMembers_studioId :: Lens' ListStudioMembers Text

The studio ID.

Destructuring the Response

data ListStudioMembersResponse

See: newListStudioMembersResponse smart constructor.

Constructors

ListStudioMembersResponse'

Fields

members :: Maybe [StudioMembership]

A list of members.

nextToken :: Maybe Text

The token for the next set of results, or null if there are no more results.

httpStatus :: Int

The response's http status code.

newListStudioMembersResponse

Arguments

:: Int
$sel:httpStatus:ListStudioMembersResponse'

-> ListStudioMembersResponse

Create a value of ListStudioMembersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:members:ListStudioMembersResponse', listStudioMembersResponse_members - A list of members.

$sel:nextToken:ListStudioMembers', listStudioMembersResponse_nextToken - The token for the next set of results, or null if there are no more results.

$sel:httpStatus:ListStudioMembersResponse', listStudioMembersResponse_httpStatus - The response's http status code.

Response Lenses

listStudioMembersResponse_members :: Lens' ListStudioMembersResponse (Maybe [StudioMembership])

A list of members.

listStudioMembersResponse_nextToken :: Lens' ListStudioMembersResponse (Maybe Text)

The token for the next set of results, or null if there are no more results.

listStudioMembersResponse_httpStatus :: Lens' ListStudioMembersResponse Int

The response's http status code.