Network.AWS.Nimble.Types.ActiveDirectoryConfiguration

Description

Documentation

data ActiveDirectoryConfiguration

The configuration for a Microsoft Active Directory (Microsoft AD) studio resource.

See: newActiveDirectoryConfiguration smart constructor.

Constructors

ActiveDirectoryConfiguration'

Fields

directoryId :: Maybe Text

The directory ID of the Directory Service for Microsoft Active Directory to access using this studio component.

computerAttributes :: Maybe [ActiveDirectoryComputerAttribute]

A collection of custom attributes for an Active Directory computer.

organizationalUnitDistinguishedName :: Maybe Text

The distinguished name (DN) and organizational unit (OU) of an Active Directory computer.

newActiveDirectoryConfiguration :: ActiveDirectoryConfiguration

Create a value of ActiveDirectoryConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryId:ActiveDirectoryConfiguration', activeDirectoryConfiguration_directoryId - The directory ID of the Directory Service for Microsoft Active Directory to access using this studio component.

$sel:computerAttributes:ActiveDirectoryConfiguration', activeDirectoryConfiguration_computerAttributes - A collection of custom attributes for an Active Directory computer.

$sel:organizationalUnitDistinguishedName:ActiveDirectoryConfiguration', activeDirectoryConfiguration_organizationalUnitDistinguishedName - The distinguished name (DN) and organizational unit (OU) of an Active Directory computer.

activeDirectoryConfiguration_directoryId :: Lens' ActiveDirectoryConfiguration (Maybe Text)

The directory ID of the Directory Service for Microsoft Active Directory to access using this studio component.

activeDirectoryConfiguration_computerAttributes :: Lens' ActiveDirectoryConfiguration (Maybe [ActiveDirectoryComputerAttribute])

A collection of custom attributes for an Active Directory computer.

activeDirectoryConfiguration_organizationalUnitDistinguishedName :: Lens' ActiveDirectoryConfiguration (Maybe Text)

The distinguished name (DN) and organizational unit (OU) of an Active Directory computer.