Network.AWS.Nimble.Types.LaunchProfileInitialization

Description

Documentation

data LaunchProfileInitialization

See: newLaunchProfileInitialization smart constructor.

Constructors

LaunchProfileInitialization'

Fields

platform :: Maybe LaunchProfilePlatform

The platform of the launch platform, either WINDOWS or LINUX.

activeDirectory :: Maybe LaunchProfileInitializationActiveDirectory

A LaunchProfileInitializationActiveDirectory resource.

launchPurpose :: Maybe Text

The launch purpose.

launchProfileId :: Maybe Text

The launch profile ID.

ec2SecurityGroupIds :: Maybe (NonEmpty Text)

The EC2 security groups that control access to the studio component.

name :: Maybe Text

The name for the launch profile.

launchProfileProtocolVersion :: Maybe Text

The version number of the protocol that is used by the launch profile. The only valid version is "2021-03-31".

userInitializationScripts :: Maybe [LaunchProfileInitializationScript]

The user initializtion scripts.

systemInitializationScripts :: Maybe [LaunchProfileInitializationScript]

The system initializtion scripts.

newLaunchProfileInitialization :: LaunchProfileInitialization

Create a value of LaunchProfileInitialization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:platform:LaunchProfileInitialization', launchProfileInitialization_platform - The platform of the launch platform, either WINDOWS or LINUX.

$sel:activeDirectory:LaunchProfileInitialization', launchProfileInitialization_activeDirectory - A LaunchProfileInitializationActiveDirectory resource.

$sel:launchPurpose:LaunchProfileInitialization', launchProfileInitialization_launchPurpose - The launch purpose.

$sel:launchProfileId:LaunchProfileInitialization', launchProfileInitialization_launchProfileId - The launch profile ID.

$sel:ec2SecurityGroupIds:LaunchProfileInitialization', launchProfileInitialization_ec2SecurityGroupIds - The EC2 security groups that control access to the studio component.

$sel:name:LaunchProfileInitialization', launchProfileInitialization_name - The name for the launch profile.

$sel:launchProfileProtocolVersion:LaunchProfileInitialization', launchProfileInitialization_launchProfileProtocolVersion - The version number of the protocol that is used by the launch profile. The only valid version is "2021-03-31".

$sel:userInitializationScripts:LaunchProfileInitialization', launchProfileInitialization_userInitializationScripts - The user initializtion scripts.

$sel:systemInitializationScripts:LaunchProfileInitialization', launchProfileInitialization_systemInitializationScripts - The system initializtion scripts.

launchProfileInitialization_platform :: Lens' LaunchProfileInitialization (Maybe LaunchProfilePlatform)

The platform of the launch platform, either WINDOWS or LINUX.

launchProfileInitialization_activeDirectory :: Lens' LaunchProfileInitialization (Maybe LaunchProfileInitializationActiveDirectory)

A LaunchProfileInitializationActiveDirectory resource.

launchProfileInitialization_launchPurpose :: Lens' LaunchProfileInitialization (Maybe Text)

The launch purpose.

launchProfileInitialization_launchProfileId :: Lens' LaunchProfileInitialization (Maybe Text)

The launch profile ID.

launchProfileInitialization_ec2SecurityGroupIds :: Lens' LaunchProfileInitialization (Maybe (NonEmpty Text))

The EC2 security groups that control access to the studio component.

launchProfileInitialization_name :: Lens' LaunchProfileInitialization (Maybe Text)

The name for the launch profile.

launchProfileInitialization_launchProfileProtocolVersion :: Lens' LaunchProfileInitialization (Maybe Text)

The version number of the protocol that is used by the launch profile. The only valid version is "2021-03-31".

launchProfileInitialization_userInitializationScripts :: Lens' LaunchProfileInitialization (Maybe [LaunchProfileInitializationScript])

The user initializtion scripts.

launchProfileInitialization_systemInitializationScripts :: Lens' LaunchProfileInitialization (Maybe [LaunchProfileInitializationScript])

The system initializtion scripts.