Network.AWS.Nimble.Types.LaunchProfileInitializationActiveDirectory

Description

Documentation

data LaunchProfileInitializationActiveDirectory

See: newLaunchProfileInitializationActiveDirectory smart constructor.

Constructors

LaunchProfileInitializationActiveDirectory'

Fields

directoryId :: Maybe Text

The directory ID of the Directory Service for Microsoft Active Directory to access using this launch profile.

studioComponentId :: Maybe Text

The unique identifier for a studio component resource.

studioComponentName :: Maybe Text

The name for the studio component.

dnsIpAddresses :: Maybe [Text]

The DNS IP address.

computerAttributes :: Maybe [ActiveDirectoryComputerAttribute]

A collection of custom attributes for an Active Directory computer.

organizationalUnitDistinguishedName :: Maybe Text

The name for the organizational unit distinguished name.

directoryName :: Maybe Text

The directory name.

newLaunchProfileInitializationActiveDirectory :: LaunchProfileInitializationActiveDirectory

Create a value of LaunchProfileInitializationActiveDirectory with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:directoryId:LaunchProfileInitializationActiveDirectory', launchProfileInitializationActiveDirectory_directoryId - The directory ID of the Directory Service for Microsoft Active Directory to access using this launch profile.

$sel:studioComponentId:LaunchProfileInitializationActiveDirectory', launchProfileInitializationActiveDirectory_studioComponentId - The unique identifier for a studio component resource.

$sel:studioComponentName:LaunchProfileInitializationActiveDirectory', launchProfileInitializationActiveDirectory_studioComponentName - The name for the studio component.

$sel:dnsIpAddresses:LaunchProfileInitializationActiveDirectory', launchProfileInitializationActiveDirectory_dnsIpAddresses - The DNS IP address.

$sel:computerAttributes:LaunchProfileInitializationActiveDirectory', launchProfileInitializationActiveDirectory_computerAttributes - A collection of custom attributes for an Active Directory computer.

$sel:organizationalUnitDistinguishedName:LaunchProfileInitializationActiveDirectory', launchProfileInitializationActiveDirectory_organizationalUnitDistinguishedName - The name for the organizational unit distinguished name.

$sel:directoryName:LaunchProfileInitializationActiveDirectory', launchProfileInitializationActiveDirectory_directoryName - The directory name.

launchProfileInitializationActiveDirectory_directoryId :: Lens' LaunchProfileInitializationActiveDirectory (Maybe Text)

The directory ID of the Directory Service for Microsoft Active Directory to access using this launch profile.

launchProfileInitializationActiveDirectory_studioComponentId :: Lens' LaunchProfileInitializationActiveDirectory (Maybe Text)

The unique identifier for a studio component resource.

launchProfileInitializationActiveDirectory_studioComponentName :: Lens' LaunchProfileInitializationActiveDirectory (Maybe Text)

The name for the studio component.

launchProfileInitializationActiveDirectory_dnsIpAddresses :: Lens' LaunchProfileInitializationActiveDirectory (Maybe [Text])

The DNS IP address.

launchProfileInitializationActiveDirectory_computerAttributes :: Lens' LaunchProfileInitializationActiveDirectory (Maybe [ActiveDirectoryComputerAttribute])

A collection of custom attributes for an Active Directory computer.

launchProfileInitializationActiveDirectory_organizationalUnitDistinguishedName :: Lens' LaunchProfileInitializationActiveDirectory (Maybe Text)

The name for the organizational unit distinguished name.

launchProfileInitializationActiveDirectory_directoryName :: Lens' LaunchProfileInitializationActiveDirectory (Maybe Text)

The directory name.