Network.AWS.Nimble.Types.StreamConfigurationCreate

Description

Documentation

data StreamConfigurationCreate

See: newStreamConfigurationCreate smart constructor.

Constructors

StreamConfigurationCreate'

Fields

maxSessionLengthInMinutes :: Maybe Natural

The length of time, in minutes, that a streaming session can run. After this point, Nimble Studio automatically terminates the session.

clipboardMode :: StreamingClipboardMode

Enable or disable the use of the system clipboard to copy and paste between the streaming session and streaming client.

streamingImageIds :: [Text]

The streaming images that users can select from when launching a streaming session with this launch profile.

ec2InstanceTypes :: NonEmpty StreamingInstanceType

The EC2 instance types that users can select from when launching a streaming session with this launch profile.

newStreamConfigurationCreate

Arguments

:: StreamingClipboardMode
$sel:clipboardMode:StreamConfigurationCreate'

-> NonEmpty StreamingInstanceType
$sel:ec2InstanceTypes:StreamConfigurationCreate'

-> StreamConfigurationCreate

Create a value of StreamConfigurationCreate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxSessionLengthInMinutes:StreamConfigurationCreate', streamConfigurationCreate_maxSessionLengthInMinutes - The length of time, in minutes, that a streaming session can run. After this point, Nimble Studio automatically terminates the session.

$sel:clipboardMode:StreamConfigurationCreate', streamConfigurationCreate_clipboardMode - Enable or disable the use of the system clipboard to copy and paste between the streaming session and streaming client.

$sel:streamingImageIds:StreamConfigurationCreate', streamConfigurationCreate_streamingImageIds - The streaming images that users can select from when launching a streaming session with this launch profile.

$sel:ec2InstanceTypes:StreamConfigurationCreate', streamConfigurationCreate_ec2InstanceTypes - The EC2 instance types that users can select from when launching a streaming session with this launch profile.

streamConfigurationCreate_maxSessionLengthInMinutes :: Lens' StreamConfigurationCreate (Maybe Natural)

The length of time, in minutes, that a streaming session can run. After this point, Nimble Studio automatically terminates the session.

streamConfigurationCreate_clipboardMode :: Lens' StreamConfigurationCreate StreamingClipboardMode

Enable or disable the use of the system clipboard to copy and paste between the streaming session and streaming client.

streamConfigurationCreate_streamingImageIds :: Lens' StreamConfigurationCreate [Text]

The streaming images that users can select from when launching a streaming session with this launch profile.

streamConfigurationCreate_ec2InstanceTypes :: Lens' StreamConfigurationCreate (NonEmpty StreamingInstanceType)

The EC2 instance types that users can select from when launching a streaming session with this launch profile.