Network.AWS.Nimble.Types.StreamingSession

Description

Documentation

data StreamingSession

See: newStreamingSession smart constructor.

Constructors

StreamingSession'

Fields

ownedBy :: Maybe Text

The user ID of the user that owns the streaming session.

state :: Maybe StreamingSessionState

The current state.

arn :: Maybe Text

The ARN of the resource.

createdAt :: Maybe POSIX

The Unix epoch timestamp in seconds for when the resource was created.

ec2InstanceType :: Maybe Text

The EC2 Instance type used for the streaming session.

createdBy :: Maybe Text

The user ID of the user that created the streaming session.

launchProfileId :: Maybe Text

The ID of the launch profile used to control access from the streaming session.

streamingImageId :: Maybe Text

The ID of the streaming image.

updatedBy :: Maybe Text

The user ID of the user that most recently updated the resource.

terminateAt :: Maybe POSIX

The time the streaming session will automatically terminate if not terminated by the user.

statusMessage :: Maybe Text

The status message for the streaming session.

updatedAt :: Maybe POSIX

The Unix epoch timestamp in seconds for when the resource was updated.

sessionId :: Maybe Text

The session ID.

tags :: Maybe (HashMap Text Text)

A collection of labels, in the form of key:value pairs, that apply to this resource.

statusCode :: Maybe StreamingSessionStatusCode

The status code.

newStreamingSession :: StreamingSession

Create a value of StreamingSession with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownedBy:StreamingSession', streamingSession_ownedBy - The user ID of the user that owns the streaming session.

$sel:state:StreamingSession', streamingSession_state - The current state.

$sel:arn:StreamingSession', streamingSession_arn - The ARN of the resource.

$sel:createdAt:StreamingSession', streamingSession_createdAt - The Unix epoch timestamp in seconds for when the resource was created.

$sel:ec2InstanceType:StreamingSession', streamingSession_ec2InstanceType - The EC2 Instance type used for the streaming session.

$sel:createdBy:StreamingSession', streamingSession_createdBy - The user ID of the user that created the streaming session.

$sel:launchProfileId:StreamingSession', streamingSession_launchProfileId - The ID of the launch profile used to control access from the streaming session.

$sel:streamingImageId:StreamingSession', streamingSession_streamingImageId - The ID of the streaming image.

$sel:updatedBy:StreamingSession', streamingSession_updatedBy - The user ID of the user that most recently updated the resource.

$sel:terminateAt:StreamingSession', streamingSession_terminateAt - The time the streaming session will automatically terminate if not terminated by the user.

$sel:statusMessage:StreamingSession', streamingSession_statusMessage - The status message for the streaming session.

$sel:updatedAt:StreamingSession', streamingSession_updatedAt - The Unix epoch timestamp in seconds for when the resource was updated.

$sel:sessionId:StreamingSession', streamingSession_sessionId - The session ID.

$sel:tags:StreamingSession', streamingSession_tags - A collection of labels, in the form of key:value pairs, that apply to this resource.

$sel:statusCode:StreamingSession', streamingSession_statusCode - The status code.

streamingSession_ownedBy :: Lens' StreamingSession (Maybe Text)

The user ID of the user that owns the streaming session.

streamingSession_state :: Lens' StreamingSession (Maybe StreamingSessionState)

The current state.

streamingSession_arn :: Lens' StreamingSession (Maybe Text)

The ARN of the resource.

streamingSession_createdAt :: Lens' StreamingSession (Maybe UTCTime)

The Unix epoch timestamp in seconds for when the resource was created.

streamingSession_ec2InstanceType :: Lens' StreamingSession (Maybe Text)

The EC2 Instance type used for the streaming session.

streamingSession_createdBy :: Lens' StreamingSession (Maybe Text)

The user ID of the user that created the streaming session.

streamingSession_launchProfileId :: Lens' StreamingSession (Maybe Text)

The ID of the launch profile used to control access from the streaming session.

streamingSession_streamingImageId :: Lens' StreamingSession (Maybe Text)

The ID of the streaming image.

streamingSession_updatedBy :: Lens' StreamingSession (Maybe Text)

The user ID of the user that most recently updated the resource.

streamingSession_terminateAt :: Lens' StreamingSession (Maybe UTCTime)

The time the streaming session will automatically terminate if not terminated by the user.

streamingSession_statusMessage :: Lens' StreamingSession (Maybe Text)

The status message for the streaming session.

streamingSession_updatedAt :: Lens' StreamingSession (Maybe UTCTime)

The Unix epoch timestamp in seconds for when the resource was updated.

streamingSession_sessionId :: Lens' StreamingSession (Maybe Text)

The session ID.

streamingSession_tags :: Lens' StreamingSession (Maybe (HashMap Text Text))

A collection of labels, in the form of key:value pairs, that apply to this resource.

streamingSession_statusCode :: Lens' StreamingSession (Maybe StreamingSessionStatusCode)

The status code.