Network.AWS.Nimble.Types.Studio

Description

Documentation

data Studio

See: newStudio smart constructor.

Constructors

Studio'

Fields

studioEncryptionConfiguration :: Maybe StudioEncryptionConfiguration

Configuration of the encryption method that is used for the studio.

state :: Maybe StudioState

The current state of the studio resource.

studioName :: Maybe Text

The name of the studio, as included in the URL when accessing it in the Nimble Studio portal.

arn :: Maybe Text

The Amazon Resource Name (ARN) that is assigned to a studio resource and uniquely identifies it. ARNs are unique across all Regions.

createdAt :: Maybe POSIX

The Unix epoch timestamp in seconds for when the resource was created.

studioId :: Maybe Text

The unique identifier for a studio resource. In Nimble Studio, all other resources are contained in a studio resource.

userRoleArn :: Maybe Text

The IAM role that studio users assume when logging in to the Nimble Studio portal.

ssoClientId :: Maybe Text

The Amazon Web Services SSO application client ID used to integrate with Amazon Web Services SSO to enable Amazon Web Services SSO users to log in to Nimble portal.

homeRegion :: Maybe Text

The Amazon Web Services Region where the studio resource is located.

statusMessage :: Maybe Text

Additional detail on the studio state.

displayName :: Maybe Text

A friendly name for the studio.

updatedAt :: Maybe POSIX

The Unix epoch timestamp in seconds for when the resource was updated.

studioUrl :: Maybe Text

The address of the web page for the studio.

adminRoleArn :: Maybe Text

The IAM role that studio admins assume when logging in to the Nimble Studio portal.

tags :: Maybe (HashMap Text Text)

A collection of labels, in the form of key:value pairs, that apply to this resource.

statusCode :: Maybe StudioStatusCode

Status codes that provide additional detail on the studio state.

newStudio :: Studio

Create a value of Studio with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:studioEncryptionConfiguration:Studio', studio_studioEncryptionConfiguration - Configuration of the encryption method that is used for the studio.

$sel:state:Studio', studio_state - The current state of the studio resource.

$sel:studioName:Studio', studio_studioName - The name of the studio, as included in the URL when accessing it in the Nimble Studio portal.

$sel:arn:Studio', studio_arn - The Amazon Resource Name (ARN) that is assigned to a studio resource and uniquely identifies it. ARNs are unique across all Regions.

$sel:createdAt:Studio', studio_createdAt - The Unix epoch timestamp in seconds for when the resource was created.

$sel:studioId:Studio', studio_studioId - The unique identifier for a studio resource. In Nimble Studio, all other resources are contained in a studio resource.

$sel:userRoleArn:Studio', studio_userRoleArn - The IAM role that studio users assume when logging in to the Nimble Studio portal.

$sel:ssoClientId:Studio', studio_ssoClientId - The Amazon Web Services SSO application client ID used to integrate with Amazon Web Services SSO to enable Amazon Web Services SSO users to log in to Nimble portal.

$sel:homeRegion:Studio', studio_homeRegion - The Amazon Web Services Region where the studio resource is located.

$sel:statusMessage:Studio', studio_statusMessage - Additional detail on the studio state.

$sel:displayName:Studio', studio_displayName - A friendly name for the studio.

$sel:updatedAt:Studio', studio_updatedAt - The Unix epoch timestamp in seconds for when the resource was updated.

$sel:studioUrl:Studio', studio_studioUrl - The address of the web page for the studio.

$sel:adminRoleArn:Studio', studio_adminRoleArn - The IAM role that studio admins assume when logging in to the Nimble Studio portal.

$sel:tags:Studio', studio_tags - A collection of labels, in the form of key:value pairs, that apply to this resource.

$sel:statusCode:Studio', studio_statusCode - Status codes that provide additional detail on the studio state.

studio_studioEncryptionConfiguration :: Lens' Studio (Maybe StudioEncryptionConfiguration)

Configuration of the encryption method that is used for the studio.

studio_state :: Lens' Studio (Maybe StudioState)

The current state of the studio resource.

studio_studioName :: Lens' Studio (Maybe Text)

The name of the studio, as included in the URL when accessing it in the Nimble Studio portal.

studio_arn :: Lens' Studio (Maybe Text)

The Amazon Resource Name (ARN) that is assigned to a studio resource and uniquely identifies it. ARNs are unique across all Regions.

studio_createdAt :: Lens' Studio (Maybe UTCTime)

The Unix epoch timestamp in seconds for when the resource was created.

studio_studioId :: Lens' Studio (Maybe Text)

The unique identifier for a studio resource. In Nimble Studio, all other resources are contained in a studio resource.

studio_userRoleArn :: Lens' Studio (Maybe Text)

The IAM role that studio users assume when logging in to the Nimble Studio portal.

studio_ssoClientId :: Lens' Studio (Maybe Text)

The Amazon Web Services SSO application client ID used to integrate with Amazon Web Services SSO to enable Amazon Web Services SSO users to log in to Nimble portal.

studio_homeRegion :: Lens' Studio (Maybe Text)

The Amazon Web Services Region where the studio resource is located.

studio_statusMessage :: Lens' Studio (Maybe Text)

Additional detail on the studio state.

studio_displayName :: Lens' Studio (Maybe Text)

A friendly name for the studio.

studio_updatedAt :: Lens' Studio (Maybe UTCTime)

The Unix epoch timestamp in seconds for when the resource was updated.

studio_studioUrl :: Lens' Studio (Maybe Text)

The address of the web page for the studio.

studio_adminRoleArn :: Lens' Studio (Maybe Text)

The IAM role that studio admins assume when logging in to the Nimble Studio portal.

studio_tags :: Lens' Studio (Maybe (HashMap Text Text))

A collection of labels, in the form of key:value pairs, that apply to this resource.

studio_statusCode :: Lens' Studio (Maybe StudioStatusCode)

Status codes that provide additional detail on the studio state.