Network.AWS.Nimble.Types.StudioComponent

Description

Documentation

data StudioComponent

A network that is used by a studio’s users and workflows, including render farm, Active Directory, licensing, and file system.

See: newStudioComponent smart constructor.

Constructors

StudioComponent'

Fields

initializationScripts :: Maybe [StudioComponentInitializationScript]

Initialization scripts for studio components.

state :: Maybe StudioComponentState

The current state.

studioComponentId :: Maybe Text

The unique identifier for a studio component resource.

arn :: Maybe Text

The ARN of the resource.

createdAt :: Maybe POSIX

The Unix epoch timestamp in seconds for when the resource was created.

createdBy :: Maybe Text

The user ID of the user that created the studio component.

ec2SecurityGroupIds :: Maybe (NonEmpty Text)

The EC2 security groups that control access to the studio component.

updatedBy :: Maybe Text

The user ID of the user that most recently updated the resource.

subtype :: Maybe StudioComponentSubtype

The specific subtype of a studio component.

name :: Maybe Text

A friendly name for the studio component resource.

statusMessage :: Maybe Text

The status message for the studio component.

scriptParameters :: Maybe [ScriptParameterKeyValue]

Parameters for the studio component scripts.

updatedAt :: Maybe POSIX

The Unix epoch timestamp in seconds for when the resource was updated.

type' :: Maybe StudioComponentType

The type of the studio component.

configuration :: Maybe StudioComponentConfiguration

The configuration of the studio component, based on component type.

description :: Maybe Text

A human-readable description for the studio component resource.

tags :: Maybe (HashMap Text Text)

A collection of labels, in the form of key:value pairs, that apply to this resource.

statusCode :: Maybe StudioComponentStatusCode

The status code.

newStudioComponent :: StudioComponent

Create a value of StudioComponent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:initializationScripts:StudioComponent', studioComponent_initializationScripts - Initialization scripts for studio components.

$sel:state:StudioComponent', studioComponent_state - The current state.

$sel:studioComponentId:StudioComponent', studioComponent_studioComponentId - The unique identifier for a studio component resource.

$sel:arn:StudioComponent', studioComponent_arn - The ARN of the resource.

$sel:createdAt:StudioComponent', studioComponent_createdAt - The Unix epoch timestamp in seconds for when the resource was created.

$sel:createdBy:StudioComponent', studioComponent_createdBy - The user ID of the user that created the studio component.

$sel:ec2SecurityGroupIds:StudioComponent', studioComponent_ec2SecurityGroupIds - The EC2 security groups that control access to the studio component.

$sel:updatedBy:StudioComponent', studioComponent_updatedBy - The user ID of the user that most recently updated the resource.

$sel:subtype:StudioComponent', studioComponent_subtype - The specific subtype of a studio component.

$sel:name:StudioComponent', studioComponent_name - A friendly name for the studio component resource.

$sel:statusMessage:StudioComponent', studioComponent_statusMessage - The status message for the studio component.

$sel:scriptParameters:StudioComponent', studioComponent_scriptParameters - Parameters for the studio component scripts.

$sel:updatedAt:StudioComponent', studioComponent_updatedAt - The Unix epoch timestamp in seconds for when the resource was updated.

$sel:type':StudioComponent', studioComponent_type - The type of the studio component.

$sel:configuration:StudioComponent', studioComponent_configuration - The configuration of the studio component, based on component type.

$sel:description:StudioComponent', studioComponent_description - A human-readable description for the studio component resource.

$sel:tags:StudioComponent', studioComponent_tags - A collection of labels, in the form of key:value pairs, that apply to this resource.

$sel:statusCode:StudioComponent', studioComponent_statusCode - The status code.

studioComponent_initializationScripts :: Lens' StudioComponent (Maybe [StudioComponentInitializationScript])

Initialization scripts for studio components.

studioComponent_state :: Lens' StudioComponent (Maybe StudioComponentState)

The current state.

studioComponent_studioComponentId :: Lens' StudioComponent (Maybe Text)

The unique identifier for a studio component resource.

studioComponent_arn :: Lens' StudioComponent (Maybe Text)

The ARN of the resource.

studioComponent_createdAt :: Lens' StudioComponent (Maybe UTCTime)

The Unix epoch timestamp in seconds for when the resource was created.

studioComponent_createdBy :: Lens' StudioComponent (Maybe Text)

The user ID of the user that created the studio component.

studioComponent_ec2SecurityGroupIds :: Lens' StudioComponent (Maybe (NonEmpty Text))

The EC2 security groups that control access to the studio component.

studioComponent_updatedBy :: Lens' StudioComponent (Maybe Text)

The user ID of the user that most recently updated the resource.

studioComponent_subtype :: Lens' StudioComponent (Maybe StudioComponentSubtype)

The specific subtype of a studio component.

studioComponent_name :: Lens' StudioComponent (Maybe Text)

A friendly name for the studio component resource.

studioComponent_statusMessage :: Lens' StudioComponent (Maybe Text)

The status message for the studio component.

studioComponent_scriptParameters :: Lens' StudioComponent (Maybe [ScriptParameterKeyValue])

Parameters for the studio component scripts.

studioComponent_updatedAt :: Lens' StudioComponent (Maybe UTCTime)

The Unix epoch timestamp in seconds for when the resource was updated.

studioComponent_type :: Lens' StudioComponent (Maybe StudioComponentType)

The type of the studio component.

studioComponent_configuration :: Lens' StudioComponent (Maybe StudioComponentConfiguration)

The configuration of the studio component, based on component type.

studioComponent_description :: Lens' StudioComponent (Maybe Text)

A human-readable description for the studio component resource.

studioComponent_tags :: Lens' StudioComponent (Maybe (HashMap Text Text))

A collection of labels, in the form of key:value pairs, that apply to this resource.

studioComponent_statusCode :: Lens' StudioComponent (Maybe StudioComponentStatusCode)

The status code.