Network.AWS.Nimble.Types.StudioComponentConfiguration

Description

Documentation

data StudioComponentConfiguration

The configuration of the studio component, based on component type.

See: newStudioComponentConfiguration smart constructor.

Constructors

StudioComponentConfiguration'

Fields

activeDirectoryConfiguration :: Maybe ActiveDirectoryConfiguration

The configuration for a Microsoft Active Directory (Microsoft AD) studio resource.

licenseServiceConfiguration :: Maybe LicenseServiceConfiguration

The configuration for a license service that is associated with a studio resource.

sharedFileSystemConfiguration :: Maybe SharedFileSystemConfiguration

The configuration for a shared file storage system that is associated with a studio resource.

computeFarmConfiguration :: Maybe ComputeFarmConfiguration

The configuration for a render farm that is associated with a studio resource.

newStudioComponentConfiguration :: StudioComponentConfiguration

Create a value of StudioComponentConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:activeDirectoryConfiguration:StudioComponentConfiguration', studioComponentConfiguration_activeDirectoryConfiguration - The configuration for a Microsoft Active Directory (Microsoft AD) studio resource.

$sel:licenseServiceConfiguration:StudioComponentConfiguration', studioComponentConfiguration_licenseServiceConfiguration - The configuration for a license service that is associated with a studio resource.

$sel:sharedFileSystemConfiguration:StudioComponentConfiguration', studioComponentConfiguration_sharedFileSystemConfiguration - The configuration for a shared file storage system that is associated with a studio resource.

$sel:computeFarmConfiguration:StudioComponentConfiguration', studioComponentConfiguration_computeFarmConfiguration - The configuration for a render farm that is associated with a studio resource.

studioComponentConfiguration_activeDirectoryConfiguration :: Lens' StudioComponentConfiguration (Maybe ActiveDirectoryConfiguration)

The configuration for a Microsoft Active Directory (Microsoft AD) studio resource.

studioComponentConfiguration_licenseServiceConfiguration :: Lens' StudioComponentConfiguration (Maybe LicenseServiceConfiguration)

The configuration for a license service that is associated with a studio resource.

studioComponentConfiguration_sharedFileSystemConfiguration :: Lens' StudioComponentConfiguration (Maybe SharedFileSystemConfiguration)

The configuration for a shared file storage system that is associated with a studio resource.

studioComponentConfiguration_computeFarmConfiguration :: Lens' StudioComponentConfiguration (Maybe ComputeFarmConfiguration)

The configuration for a render farm that is associated with a studio resource.