Network.AWS.Nimble.Types.StudioComponentInitializationScript

Description

Documentation

data StudioComponentInitializationScript

Initialization scripts for studio components.

See: newStudioComponentInitializationScript smart constructor.

Constructors

StudioComponentInitializationScript'

Fields

script :: Maybe Text

The initialization script.

platform :: Maybe LaunchProfilePlatform

The platform of the initialization script, either WINDOWS or LINUX.

runContext :: Maybe StudioComponentInitializationScriptRunContext

The method to use when running the initialization script.

launchProfileProtocolVersion :: Maybe Text

The version number of the protocol that is used by the launch profile. The only valid version is "2021-03-31".

newStudioComponentInitializationScript :: StudioComponentInitializationScript

Create a value of StudioComponentInitializationScript with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:script:StudioComponentInitializationScript', studioComponentInitializationScript_script - The initialization script.

$sel:platform:StudioComponentInitializationScript', studioComponentInitializationScript_platform - The platform of the initialization script, either WINDOWS or LINUX.

$sel:runContext:StudioComponentInitializationScript', studioComponentInitializationScript_runContext - The method to use when running the initialization script.

$sel:launchProfileProtocolVersion:StudioComponentInitializationScript', studioComponentInitializationScript_launchProfileProtocolVersion - The version number of the protocol that is used by the launch profile. The only valid version is "2021-03-31".

studioComponentInitializationScript_script :: Lens' StudioComponentInitializationScript (Maybe Text)

The initialization script.

studioComponentInitializationScript_platform :: Lens' StudioComponentInitializationScript (Maybe LaunchProfilePlatform)

The platform of the initialization script, either WINDOWS or LINUX.

studioComponentInitializationScript_runContext :: Lens' StudioComponentInitializationScript (Maybe StudioComponentInitializationScriptRunContext)

The method to use when running the initialization script.

studioComponentInitializationScript_launchProfileProtocolVersion :: Lens' StudioComponentInitializationScript (Maybe Text)

The version number of the protocol that is used by the launch profile. The only valid version is "2021-03-31".