Network.AWS.Nimble.Types.StudioEncryptionConfiguration

Description

Documentation

data StudioEncryptionConfiguration

Configuration of the encryption method that is used for the studio.

See: newStudioEncryptionConfiguration smart constructor.

Constructors

StudioEncryptionConfiguration'

Fields

keyArn :: Maybe Text

The ARN for a KMS key that is used to encrypt studio data.

keyType :: StudioEncryptionConfigurationKeyType

The type of KMS key that is used to encrypt studio data.

newStudioEncryptionConfiguration

Arguments

:: StudioEncryptionConfigurationKeyType
$sel:keyType:StudioEncryptionConfiguration'

-> StudioEncryptionConfiguration

Create a value of StudioEncryptionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyArn:StudioEncryptionConfiguration', studioEncryptionConfiguration_keyArn - The ARN for a KMS key that is used to encrypt studio data.

$sel:keyType:StudioEncryptionConfiguration', studioEncryptionConfiguration_keyType - The type of KMS key that is used to encrypt studio data.

studioEncryptionConfiguration_keyArn :: Lens' StudioEncryptionConfiguration (Maybe Text)

The ARN for a KMS key that is used to encrypt studio data.

studioEncryptionConfiguration_keyType :: Lens' StudioEncryptionConfiguration StudioEncryptionConfigurationKeyType

The type of KMS key that is used to encrypt studio data.