Network.AWS.Nimble.UpdateLaunchProfileMember

Description

Update a user persona in launch profile membership.

Creating a Request

data UpdateLaunchProfileMember

The principal ID.

See: newUpdateLaunchProfileMember smart constructor.

Constructors

UpdateLaunchProfileMember'

Fields

clientToken :: Maybe Text

To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

studioId :: Text

The studio ID.

persona :: LaunchProfilePersona

The persona.

principalId :: Text

The principal ID.

launchProfileId :: Text

The launch profile ID.

newUpdateLaunchProfileMember

Arguments

:: Text
$sel:studioId:UpdateLaunchProfileMember'

-> LaunchProfilePersona
$sel:persona:UpdateLaunchProfileMember'

-> Text
$sel:principalId:UpdateLaunchProfileMember'

-> Text
$sel:launchProfileId:UpdateLaunchProfileMember'

-> UpdateLaunchProfileMember

Create a value of UpdateLaunchProfileMember with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:UpdateLaunchProfileMember', updateLaunchProfileMember_clientToken - To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

$sel:studioId:UpdateLaunchProfileMember', updateLaunchProfileMember_studioId - The studio ID.

$sel:persona:UpdateLaunchProfileMember', updateLaunchProfileMember_persona - The persona.

$sel:principalId:UpdateLaunchProfileMember', updateLaunchProfileMember_principalId - The principal ID.

$sel:launchProfileId:UpdateLaunchProfileMember', updateLaunchProfileMember_launchProfileId - The launch profile ID.

Request Lenses

updateLaunchProfileMember_clientToken :: Lens' UpdateLaunchProfileMember (Maybe Text)

To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

updateLaunchProfileMember_studioId :: Lens' UpdateLaunchProfileMember Text

The studio ID.

updateLaunchProfileMember_persona :: Lens' UpdateLaunchProfileMember LaunchProfilePersona

The persona.

updateLaunchProfileMember_principalId :: Lens' UpdateLaunchProfileMember Text

The principal ID.

updateLaunchProfileMember_launchProfileId :: Lens' UpdateLaunchProfileMember Text

The launch profile ID.

Destructuring the Response

data UpdateLaunchProfileMemberResponse

See: newUpdateLaunchProfileMemberResponse smart constructor.

Constructors

UpdateLaunchProfileMemberResponse'

Fields

member :: Maybe LaunchProfileMembership

The member.

httpStatus :: Int

The response's http status code.

newUpdateLaunchProfileMemberResponse

Arguments

:: Int
$sel:httpStatus:UpdateLaunchProfileMemberResponse'

-> UpdateLaunchProfileMemberResponse

Create a value of UpdateLaunchProfileMemberResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:member:UpdateLaunchProfileMemberResponse', updateLaunchProfileMemberResponse_member - The member.

$sel:httpStatus:UpdateLaunchProfileMemberResponse', updateLaunchProfileMemberResponse_httpStatus - The response's http status code.

Response Lenses

updateLaunchProfileMemberResponse_member :: Lens' UpdateLaunchProfileMemberResponse (Maybe LaunchProfileMembership)

The member.

updateLaunchProfileMemberResponse_httpStatus :: Lens' UpdateLaunchProfileMemberResponse Int

The response's http status code.