Network.AWS.Nimble.UpdateStreamingImage

Description

Update streaming image.

Creating a Request

data UpdateStreamingImage

The streaming image ID.

See: newUpdateStreamingImage smart constructor.

Constructors

UpdateStreamingImage'

Fields

clientToken :: Maybe Text

To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

name :: Maybe Text

The name for the streaming image.

description :: Maybe Text

The description.

studioId :: Text

The studio ID.

streamingImageId :: Text

The streaming image ID.

newUpdateStreamingImage

Arguments

:: Text
$sel:studioId:UpdateStreamingImage'

-> Text
$sel:streamingImageId:UpdateStreamingImage'

-> UpdateStreamingImage

Create a value of UpdateStreamingImage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:UpdateStreamingImage', updateStreamingImage_clientToken - To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

$sel:name:UpdateStreamingImage', updateStreamingImage_name - The name for the streaming image.

$sel:description:UpdateStreamingImage', updateStreamingImage_description - The description.

$sel:studioId:UpdateStreamingImage', updateStreamingImage_studioId - The studio ID.

$sel:streamingImageId:UpdateStreamingImage', updateStreamingImage_streamingImageId - The streaming image ID.

Request Lenses

updateStreamingImage_clientToken :: Lens' UpdateStreamingImage (Maybe Text)

To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

updateStreamingImage_name :: Lens' UpdateStreamingImage (Maybe Text)

The name for the streaming image.

updateStreamingImage_description :: Lens' UpdateStreamingImage (Maybe Text)

The description.

updateStreamingImage_studioId :: Lens' UpdateStreamingImage Text

The studio ID.

updateStreamingImage_streamingImageId :: Lens' UpdateStreamingImage Text

The streaming image ID.

Destructuring the Response

data UpdateStreamingImageResponse

See: newUpdateStreamingImageResponse smart constructor.

Constructors

UpdateStreamingImageResponse'

Fields

streamingImage :: Maybe StreamingImage

httpStatus :: Int

The response's http status code.

newUpdateStreamingImageResponse

Arguments

:: Int
$sel:httpStatus:UpdateStreamingImageResponse'

-> UpdateStreamingImageResponse

Create a value of UpdateStreamingImageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamingImage:UpdateStreamingImageResponse', updateStreamingImageResponse_streamingImage -

$sel:httpStatus:UpdateStreamingImageResponse', updateStreamingImageResponse_httpStatus - The response's http status code.

Response Lenses

updateStreamingImageResponse_streamingImage :: Lens' UpdateStreamingImageResponse (Maybe StreamingImage)

updateStreamingImageResponse_httpStatus :: Lens' UpdateStreamingImageResponse Int

The response's http status code.