Network.AWS.Nimble.UpdateStudioComponent

Description

Updates a studio component resource.

Creating a Request

data UpdateStudioComponent

The studio component ID.

See: newUpdateStudioComponent smart constructor.

Constructors

UpdateStudioComponent'

Fields

initializationScripts :: Maybe [StudioComponentInitializationScript]

Initialization scripts for studio components.

clientToken :: Maybe Text

To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

ec2SecurityGroupIds :: Maybe (NonEmpty Text)

The EC2 security groups that control access to the studio component.

subtype :: Maybe StudioComponentSubtype

The specific subtype of a studio component.

name :: Maybe Text

The name for the studio component.

scriptParameters :: Maybe [ScriptParameterKeyValue]

Parameters for the studio component scripts.

type' :: Maybe StudioComponentType

The type of the studio component.

configuration :: Maybe StudioComponentConfiguration

The configuration of the studio component, based on component type.

description :: Maybe Text

The description.

studioId :: Text

The studio ID.

studioComponentId :: Text

The studio component ID.

newUpdateStudioComponent

Arguments

:: Text
$sel:studioId:UpdateStudioComponent'

-> Text
$sel:studioComponentId:UpdateStudioComponent'

-> UpdateStudioComponent

Create a value of UpdateStudioComponent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:initializationScripts:UpdateStudioComponent', updateStudioComponent_initializationScripts - Initialization scripts for studio components.

$sel:clientToken:UpdateStudioComponent', updateStudioComponent_clientToken - To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

$sel:ec2SecurityGroupIds:UpdateStudioComponent', updateStudioComponent_ec2SecurityGroupIds - The EC2 security groups that control access to the studio component.

$sel:subtype:UpdateStudioComponent', updateStudioComponent_subtype - The specific subtype of a studio component.

$sel:name:UpdateStudioComponent', updateStudioComponent_name - The name for the studio component.

$sel:scriptParameters:UpdateStudioComponent', updateStudioComponent_scriptParameters - Parameters for the studio component scripts.

$sel:type':UpdateStudioComponent', updateStudioComponent_type - The type of the studio component.

$sel:configuration:UpdateStudioComponent', updateStudioComponent_configuration - The configuration of the studio component, based on component type.

$sel:description:UpdateStudioComponent', updateStudioComponent_description - The description.

$sel:studioId:UpdateStudioComponent', updateStudioComponent_studioId - The studio ID.

$sel:studioComponentId:UpdateStudioComponent', updateStudioComponent_studioComponentId - The studio component ID.

Request Lenses

updateStudioComponent_initializationScripts :: Lens' UpdateStudioComponent (Maybe [StudioComponentInitializationScript])

Initialization scripts for studio components.

updateStudioComponent_clientToken :: Lens' UpdateStudioComponent (Maybe Text)

To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

updateStudioComponent_ec2SecurityGroupIds :: Lens' UpdateStudioComponent (Maybe (NonEmpty Text))

The EC2 security groups that control access to the studio component.

updateStudioComponent_subtype :: Lens' UpdateStudioComponent (Maybe StudioComponentSubtype)

The specific subtype of a studio component.

updateStudioComponent_name :: Lens' UpdateStudioComponent (Maybe Text)

The name for the studio component.

updateStudioComponent_scriptParameters :: Lens' UpdateStudioComponent (Maybe [ScriptParameterKeyValue])

Parameters for the studio component scripts.

updateStudioComponent_type :: Lens' UpdateStudioComponent (Maybe StudioComponentType)

The type of the studio component.

updateStudioComponent_configuration :: Lens' UpdateStudioComponent (Maybe StudioComponentConfiguration)

The configuration of the studio component, based on component type.

updateStudioComponent_description :: Lens' UpdateStudioComponent (Maybe Text)

The description.

updateStudioComponent_studioId :: Lens' UpdateStudioComponent Text

The studio ID.

updateStudioComponent_studioComponentId :: Lens' UpdateStudioComponent Text

The studio component ID.

Destructuring the Response

data UpdateStudioComponentResponse

See: newUpdateStudioComponentResponse smart constructor.

Constructors

UpdateStudioComponentResponse'

Fields

studioComponent :: Maybe StudioComponent

Information about the studio component.

httpStatus :: Int

The response's http status code.

newUpdateStudioComponentResponse

Arguments

:: Int
$sel:httpStatus:UpdateStudioComponentResponse'

-> UpdateStudioComponentResponse

Create a value of UpdateStudioComponentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:studioComponent:UpdateStudioComponentResponse', updateStudioComponentResponse_studioComponent - Information about the studio component.

$sel:httpStatus:UpdateStudioComponentResponse', updateStudioComponentResponse_httpStatus - The response's http status code.

Response Lenses

updateStudioComponentResponse_studioComponent :: Lens' UpdateStudioComponentResponse (Maybe StudioComponent)

Information about the studio component.

updateStudioComponentResponse_httpStatus :: Lens' UpdateStudioComponentResponse Int

The response's http status code.