Network.AWS.OpenSearch.DeleteInboundConnection

Description

Allows the remote domain owner to delete an existing inbound cross-cluster connection.

Creating a Request

data DeleteInboundConnection

Container for the parameters to the DeleteInboundConnection operation.

See: newDeleteInboundConnection smart constructor.

Constructors

DeleteInboundConnection'

Fields

connectionId :: Text

The ID of the inbound connection to permanently delete.

newDeleteInboundConnection

Arguments

:: Text
$sel:connectionId:DeleteInboundConnection'

-> DeleteInboundConnection

Create a value of DeleteInboundConnection with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectionId:DeleteInboundConnection', deleteInboundConnection_connectionId - The ID of the inbound connection to permanently delete.

Request Lenses

deleteInboundConnection_connectionId :: Lens' DeleteInboundConnection Text

The ID of the inbound connection to permanently delete.

Destructuring the Response

data DeleteInboundConnectionResponse

The result of a DeleteInboundConnection operation. Contains details about the deleted inbound connection.

See: newDeleteInboundConnectionResponse smart constructor.

Constructors

DeleteInboundConnectionResponse'

Fields

connection :: Maybe InboundConnection

The InboundConnection of the deleted inbound connection.

httpStatus :: Int

The response's http status code.

newDeleteInboundConnectionResponse

Arguments

:: Int
$sel:httpStatus:DeleteInboundConnectionResponse'

-> DeleteInboundConnectionResponse

Create a value of DeleteInboundConnectionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connection:DeleteInboundConnectionResponse', deleteInboundConnectionResponse_connection - The InboundConnection of the deleted inbound connection.

$sel:httpStatus:DeleteInboundConnectionResponse', deleteInboundConnectionResponse_httpStatus - The response's http status code.

Response Lenses

deleteInboundConnectionResponse_connection :: Lens' DeleteInboundConnectionResponse (Maybe InboundConnection)

The InboundConnection of the deleted inbound connection.

deleteInboundConnectionResponse_httpStatus :: Lens' DeleteInboundConnectionResponse Int

The response's http status code.