Network.AWS.OpenSearch.DescribeDomainAutoTunes

Description

Provides scheduled Auto-Tune action details for the domain, such as Auto-Tune action type, description, severity, and scheduled date.

Creating a Request

data DescribeDomainAutoTunes

Container for the parameters to the DescribeDomainAutoTunes operation.

See: newDescribeDomainAutoTunes smart constructor.

Constructors

DescribeDomainAutoTunes'

Fields

nextToken :: Maybe Text

NextToken is sent in case the earlier API call results contain the NextToken. Used for pagination.

maxResults :: Maybe Int

Set this value to limit the number of results returned. If not specified, defaults to 100.

domainName :: Text

The domain name for which you want Auto-Tune action details.

newDescribeDomainAutoTunes

Arguments

:: Text
$sel:domainName:DescribeDomainAutoTunes'

-> DescribeDomainAutoTunes

Create a value of DescribeDomainAutoTunes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeDomainAutoTunes', describeDomainAutoTunes_nextToken - NextToken is sent in case the earlier API call results contain the NextToken. Used for pagination.

$sel:maxResults:DescribeDomainAutoTunes', describeDomainAutoTunes_maxResults - Set this value to limit the number of results returned. If not specified, defaults to 100.

$sel:domainName:DescribeDomainAutoTunes', describeDomainAutoTunes_domainName - The domain name for which you want Auto-Tune action details.

Request Lenses

describeDomainAutoTunes_nextToken :: Lens' DescribeDomainAutoTunes (Maybe Text)

NextToken is sent in case the earlier API call results contain the NextToken. Used for pagination.

describeDomainAutoTunes_maxResults :: Lens' DescribeDomainAutoTunes (Maybe Int)

Set this value to limit the number of results returned. If not specified, defaults to 100.

describeDomainAutoTunes_domainName :: Lens' DescribeDomainAutoTunes Text

The domain name for which you want Auto-Tune action details.

Destructuring the Response

data DescribeDomainAutoTunesResponse

The result of a DescribeDomainAutoTunes request. See Auto-Tune for Amazon OpenSearch Service for more information.

See: newDescribeDomainAutoTunesResponse smart constructor.

Constructors

DescribeDomainAutoTunesResponse'

Fields

nextToken :: Maybe Text

An identifier to allow retrieval of paginated results.

autoTunes :: Maybe [AutoTune]

The list of setting adjustments that Auto-Tune has made to the domain. See Auto-Tune for Amazon OpenSearch Service for more information.

httpStatus :: Int

The response's http status code.

newDescribeDomainAutoTunesResponse

Arguments

:: Int
$sel:httpStatus:DescribeDomainAutoTunesResponse'

-> DescribeDomainAutoTunesResponse

Create a value of DescribeDomainAutoTunesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeDomainAutoTunes', describeDomainAutoTunesResponse_nextToken - An identifier to allow retrieval of paginated results.

$sel:autoTunes:DescribeDomainAutoTunesResponse', describeDomainAutoTunesResponse_autoTunes - The list of setting adjustments that Auto-Tune has made to the domain. See Auto-Tune for Amazon OpenSearch Service for more information.

$sel:httpStatus:DescribeDomainAutoTunesResponse', describeDomainAutoTunesResponse_httpStatus - The response's http status code.

Response Lenses

describeDomainAutoTunesResponse_nextToken :: Lens' DescribeDomainAutoTunesResponse (Maybe Text)

An identifier to allow retrieval of paginated results.

describeDomainAutoTunesResponse_autoTunes :: Lens' DescribeDomainAutoTunesResponse (Maybe [AutoTune])

The list of setting adjustments that Auto-Tune has made to the domain. See Auto-Tune for Amazon OpenSearch Service for more information.

describeDomainAutoTunesResponse_httpStatus :: Lens' DescribeDomainAutoTunesResponse Int

The response's http status code.