Network.AWS.OpenSearch.DescribeReservedInstances

Description

Returns information about reserved OpenSearch instances for this account.

Creating a Request

data DescribeReservedInstances

Container for parameters to DescribeReservedInstances

See: newDescribeReservedInstances smart constructor.

Constructors

DescribeReservedInstances'

Fields

nextToken :: Maybe Text

Provides an identifier to allow retrieval of paginated results.

reservedInstanceId :: Maybe Text

The reserved instance identifier filter value. Use this parameter to show only the reservation that matches the specified reserved OpenSearch instance ID.

maxResults :: Maybe Int

Set this value to limit the number of results returned. If not specified, defaults to 100.

newDescribeReservedInstances :: DescribeReservedInstances

Create a value of DescribeReservedInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeReservedInstances', describeReservedInstances_nextToken - Provides an identifier to allow retrieval of paginated results.

$sel:reservedInstanceId:DescribeReservedInstances', describeReservedInstances_reservedInstanceId - The reserved instance identifier filter value. Use this parameter to show only the reservation that matches the specified reserved OpenSearch instance ID.

$sel:maxResults:DescribeReservedInstances', describeReservedInstances_maxResults - Set this value to limit the number of results returned. If not specified, defaults to 100.

Request Lenses

describeReservedInstances_nextToken :: Lens' DescribeReservedInstances (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

describeReservedInstances_reservedInstanceId :: Lens' DescribeReservedInstances (Maybe Text)

The reserved instance identifier filter value. Use this parameter to show only the reservation that matches the specified reserved OpenSearch instance ID.

describeReservedInstances_maxResults :: Lens' DescribeReservedInstances (Maybe Int)

Set this value to limit the number of results returned. If not specified, defaults to 100.

Destructuring the Response

data DescribeReservedInstancesResponse

Container for results from DescribeReservedInstances

See: newDescribeReservedInstancesResponse smart constructor.

Constructors

DescribeReservedInstancesResponse'

Fields

nextToken :: Maybe Text

Provides an identifier to allow retrieval of paginated results.

reservedInstances :: Maybe [ReservedInstance]

List of reserved OpenSearch instances.

httpStatus :: Int

The response's http status code.

newDescribeReservedInstancesResponse

Arguments

:: Int
$sel:httpStatus:DescribeReservedInstancesResponse'

-> DescribeReservedInstancesResponse

Create a value of DescribeReservedInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeReservedInstances', describeReservedInstancesResponse_nextToken - Provides an identifier to allow retrieval of paginated results.

$sel:reservedInstances:DescribeReservedInstancesResponse', describeReservedInstancesResponse_reservedInstances - List of reserved OpenSearch instances.

$sel:httpStatus:DescribeReservedInstancesResponse', describeReservedInstancesResponse_httpStatus - The response's http status code.

Response Lenses

describeReservedInstancesResponse_nextToken :: Lens' DescribeReservedInstancesResponse (Maybe Text)

Provides an identifier to allow retrieval of paginated results.

describeReservedInstancesResponse_reservedInstances :: Lens' DescribeReservedInstancesResponse (Maybe [ReservedInstance])

List of reserved OpenSearch instances.

describeReservedInstancesResponse_httpStatus :: Lens' DescribeReservedInstancesResponse Int

The response's http status code.