Network.AWS.OpenSearch.Types.DomainConfig

Description

Documentation

data DomainConfig

The configuration of a domain.

See: newDomainConfig smart constructor.

Constructors

DomainConfig'

Fields

eBSOptions :: Maybe EBSOptionsStatus

The EBSOptions for the domain.

engineVersion :: Maybe VersionStatus

String of format Elasticsearch_X.Y or OpenSearch_X.Y to specify the engine version for the OpenSearch or Elasticsearch domain.

nodeToNodeEncryptionOptions :: Maybe NodeToNodeEncryptionOptionsStatus

The NodeToNodeEncryptionOptions for the domain.

accessPolicies :: Maybe AccessPoliciesStatus

IAM access policy as a JSON-formatted string.

autoTuneOptions :: Maybe AutoTuneOptionsStatus

Specifies AutoTuneOptions for the domain.

logPublishingOptions :: Maybe LogPublishingOptionsStatus

Log publishing options for the given domain.

clusterConfig :: Maybe ClusterConfigStatus

The ClusterConfig for the domain.

advancedSecurityOptions :: Maybe AdvancedSecurityOptionsStatus

Specifies AdvancedSecurityOptions for the domain.

snapshotOptions :: Maybe SnapshotOptionsStatus

The SnapshotOptions for the domain.

cognitoOptions :: Maybe CognitoOptionsStatus

The CognitoOptions for the specified domain. For more information, see Configuring Amazon Cognito authentication for OpenSearch Dashboards.

encryptionAtRestOptions :: Maybe EncryptionAtRestOptionsStatus

The EncryptionAtRestOptions for the domain.

vPCOptions :: Maybe VPCDerivedInfoStatus

The VPCOptions for the specified domain. For more information, see Launching your Amazon OpenSearch Service domains using a VPC.

domainEndpointOptions :: Maybe DomainEndpointOptionsStatus

The DomainEndpointOptions for the domain.

advancedOptions :: Maybe AdvancedOptionsStatus

The AdvancedOptions for the domain. See Advanced options for more information.

newDomainConfig :: DomainConfig

Create a value of DomainConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eBSOptions:DomainConfig', domainConfig_eBSOptions - The EBSOptions for the domain.

$sel:engineVersion:DomainConfig', domainConfig_engineVersion - String of format Elasticsearch_X.Y or OpenSearch_X.Y to specify the engine version for the OpenSearch or Elasticsearch domain.

$sel:nodeToNodeEncryptionOptions:DomainConfig', domainConfig_nodeToNodeEncryptionOptions - The NodeToNodeEncryptionOptions for the domain.

$sel:accessPolicies:DomainConfig', domainConfig_accessPolicies - IAM access policy as a JSON-formatted string.

$sel:autoTuneOptions:DomainConfig', domainConfig_autoTuneOptions - Specifies AutoTuneOptions for the domain.

$sel:logPublishingOptions:DomainConfig', domainConfig_logPublishingOptions - Log publishing options for the given domain.

$sel:clusterConfig:DomainConfig', domainConfig_clusterConfig - The ClusterConfig for the domain.

$sel:advancedSecurityOptions:DomainConfig', domainConfig_advancedSecurityOptions - Specifies AdvancedSecurityOptions for the domain.

$sel:snapshotOptions:DomainConfig', domainConfig_snapshotOptions - The SnapshotOptions for the domain.

$sel:cognitoOptions:DomainConfig', domainConfig_cognitoOptions - The CognitoOptions for the specified domain. For more information, see Configuring Amazon Cognito authentication for OpenSearch Dashboards.

$sel:encryptionAtRestOptions:DomainConfig', domainConfig_encryptionAtRestOptions - The EncryptionAtRestOptions for the domain.

$sel:vPCOptions:DomainConfig', domainConfig_vPCOptions - The VPCOptions for the specified domain. For more information, see Launching your Amazon OpenSearch Service domains using a VPC.

$sel:domainEndpointOptions:DomainConfig', domainConfig_domainEndpointOptions - The DomainEndpointOptions for the domain.

$sel:advancedOptions:DomainConfig', domainConfig_advancedOptions - The AdvancedOptions for the domain. See Advanced options for more information.

domainConfig_eBSOptions :: Lens' DomainConfig (Maybe EBSOptionsStatus)

The EBSOptions for the domain.

domainConfig_engineVersion :: Lens' DomainConfig (Maybe VersionStatus)

String of format Elasticsearch_X.Y or OpenSearch_X.Y to specify the engine version for the OpenSearch or Elasticsearch domain.

domainConfig_nodeToNodeEncryptionOptions :: Lens' DomainConfig (Maybe NodeToNodeEncryptionOptionsStatus)

The NodeToNodeEncryptionOptions for the domain.

domainConfig_accessPolicies :: Lens' DomainConfig (Maybe AccessPoliciesStatus)

IAM access policy as a JSON-formatted string.

domainConfig_autoTuneOptions :: Lens' DomainConfig (Maybe AutoTuneOptionsStatus)

Specifies AutoTuneOptions for the domain.

domainConfig_logPublishingOptions :: Lens' DomainConfig (Maybe LogPublishingOptionsStatus)

Log publishing options for the given domain.

domainConfig_clusterConfig :: Lens' DomainConfig (Maybe ClusterConfigStatus)

The ClusterConfig for the domain.

domainConfig_advancedSecurityOptions :: Lens' DomainConfig (Maybe AdvancedSecurityOptionsStatus)

Specifies AdvancedSecurityOptions for the domain.

domainConfig_snapshotOptions :: Lens' DomainConfig (Maybe SnapshotOptionsStatus)

The SnapshotOptions for the domain.

domainConfig_cognitoOptions :: Lens' DomainConfig (Maybe CognitoOptionsStatus)

The CognitoOptions for the specified domain. For more information, see Configuring Amazon Cognito authentication for OpenSearch Dashboards.

domainConfig_encryptionAtRestOptions :: Lens' DomainConfig (Maybe EncryptionAtRestOptionsStatus)

The EncryptionAtRestOptions for the domain.

domainConfig_vPCOptions :: Lens' DomainConfig (Maybe VPCDerivedInfoStatus)

The VPCOptions for the specified domain. For more information, see Launching your Amazon OpenSearch Service domains using a VPC.

domainConfig_domainEndpointOptions :: Lens' DomainConfig (Maybe DomainEndpointOptionsStatus)

The DomainEndpointOptions for the domain.

domainConfig_advancedOptions :: Lens' DomainConfig (Maybe AdvancedOptionsStatus)

The AdvancedOptions for the domain. See Advanced options for more information.