Network.AWS.OpenSearch.Types.ReservedInstanceOffering

Description

Documentation

data ReservedInstanceOffering

Details of a reserved OpenSearch instance offering.

See: newReservedInstanceOffering smart constructor.

Constructors

ReservedInstanceOffering'

Fields

currencyCode :: Maybe Text

The currency code for the reserved OpenSearch instance offering.

instanceType :: Maybe OpenSearchPartitionInstanceType

The OpenSearch instance type offered by the reserved instance offering.

recurringCharges :: Maybe [RecurringCharge]

The charge to your account regardless of whether you are creating any domains using the instance offering.

usagePrice :: Maybe Double

The rate you are charged for each hour the domain that is using the offering is running.

reservedInstanceOfferingId :: Maybe Text

The OpenSearch reserved instance offering identifier.

fixedPrice :: Maybe Double

The upfront fixed charge you will pay to purchase the specific reserved OpenSearch instance offering.

duration :: Maybe Int

The duration, in seconds, for which the offering will reserve the OpenSearch instance.

paymentOption :: Maybe ReservedInstancePaymentOption

Payment option for the reserved OpenSearch instance offering

newReservedInstanceOffering :: ReservedInstanceOffering

Create a value of ReservedInstanceOffering with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:currencyCode:ReservedInstanceOffering', reservedInstanceOffering_currencyCode - The currency code for the reserved OpenSearch instance offering.

$sel:instanceType:ReservedInstanceOffering', reservedInstanceOffering_instanceType - The OpenSearch instance type offered by the reserved instance offering.

$sel:recurringCharges:ReservedInstanceOffering', reservedInstanceOffering_recurringCharges - The charge to your account regardless of whether you are creating any domains using the instance offering.

$sel:usagePrice:ReservedInstanceOffering', reservedInstanceOffering_usagePrice - The rate you are charged for each hour the domain that is using the offering is running.

$sel:reservedInstanceOfferingId:ReservedInstanceOffering', reservedInstanceOffering_reservedInstanceOfferingId - The OpenSearch reserved instance offering identifier.

$sel:fixedPrice:ReservedInstanceOffering', reservedInstanceOffering_fixedPrice - The upfront fixed charge you will pay to purchase the specific reserved OpenSearch instance offering.

$sel:duration:ReservedInstanceOffering', reservedInstanceOffering_duration - The duration, in seconds, for which the offering will reserve the OpenSearch instance.

$sel:paymentOption:ReservedInstanceOffering', reservedInstanceOffering_paymentOption - Payment option for the reserved OpenSearch instance offering

reservedInstanceOffering_currencyCode :: Lens' ReservedInstanceOffering (Maybe Text)

The currency code for the reserved OpenSearch instance offering.

reservedInstanceOffering_instanceType :: Lens' ReservedInstanceOffering (Maybe OpenSearchPartitionInstanceType)

The OpenSearch instance type offered by the reserved instance offering.

reservedInstanceOffering_recurringCharges :: Lens' ReservedInstanceOffering (Maybe [RecurringCharge])

The charge to your account regardless of whether you are creating any domains using the instance offering.

reservedInstanceOffering_usagePrice :: Lens' ReservedInstanceOffering (Maybe Double)

The rate you are charged for each hour the domain that is using the offering is running.

reservedInstanceOffering_reservedInstanceOfferingId :: Lens' ReservedInstanceOffering (Maybe Text)

The OpenSearch reserved instance offering identifier.

reservedInstanceOffering_fixedPrice :: Lens' ReservedInstanceOffering (Maybe Double)

The upfront fixed charge you will pay to purchase the specific reserved OpenSearch instance offering.

reservedInstanceOffering_duration :: Lens' ReservedInstanceOffering (Maybe Int)

The duration, in seconds, for which the offering will reserve the OpenSearch instance.

reservedInstanceOffering_paymentOption :: Lens' ReservedInstanceOffering (Maybe ReservedInstancePaymentOption)

Payment option for the reserved OpenSearch instance offering