Network.AWS.OpsWorks.CreateApp

Description

Creates an app for a specified stack. For more information, see Creating Apps.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Creating a Request

data CreateApp

See: newCreateApp smart constructor.

Constructors

CreateApp'

Fields

sslConfiguration :: Maybe SslConfiguration

An SslConfiguration object with the SSL configuration.

environment :: Maybe [EnvironmentVariable]

An array of EnvironmentVariable objects that specify environment variables to be associated with the app. After you deploy the app, these variables are defined on the associated app server instance. For more information, see Environment Variables.

There is no specific limit on the number of environment variables. However, the size of the associated data structure - which includes the variables' names, values, and protected flag values - cannot exceed 20 KB. This limit should accommodate most if not all use cases. Exceeding it will cause an exception with the message, "Environment: is too large (maximum is 20KB)."

If you have specified one or more environment variables, you cannot modify the stack's Chef version.

enableSsl :: Maybe Bool

Whether to enable SSL for the app.

shortname :: Maybe Text

The app's short name.

dataSources :: Maybe [DataSource]

The app's data source.

appSource :: Maybe Source

A Source object that specifies the app repository.

attributes :: Maybe (HashMap AppAttributesKeys Text)

One or more user-defined key/value pairs to be added to the stack attributes.

domains :: Maybe [Text]

The app virtual host settings, with multiple domains separated by commas. For example: 'www.example.com, example.com'

description :: Maybe Text

A description of the app.

stackId :: Text

The stack ID.

name :: Text

The app name.

type' :: AppType

The app type. Each supported type is associated with a particular layer. For example, PHP applications are associated with a PHP layer. AWS OpsWorks Stacks deploys an application to those instances that are members of the corresponding layer. If your app isn't one of the standard types, or you prefer to implement your own Deploy recipes, specify other.

newCreateApp

Arguments

:: Text
$sel:stackId:CreateApp'

-> Text
$sel:name:CreateApp'

-> AppType
$sel:type':CreateApp'

-> CreateApp

Create a value of CreateApp with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sslConfiguration:CreateApp', createApp_sslConfiguration - An SslConfiguration object with the SSL configuration.

$sel:environment:CreateApp', createApp_environment - An array of EnvironmentVariable objects that specify environment variables to be associated with the app. After you deploy the app, these variables are defined on the associated app server instance. For more information, see Environment Variables.

There is no specific limit on the number of environment variables. However, the size of the associated data structure - which includes the variables' names, values, and protected flag values - cannot exceed 20 KB. This limit should accommodate most if not all use cases. Exceeding it will cause an exception with the message, "Environment: is too large (maximum is 20KB)."

If you have specified one or more environment variables, you cannot modify the stack's Chef version.

$sel:enableSsl:CreateApp', createApp_enableSsl - Whether to enable SSL for the app.

$sel:shortname:CreateApp', createApp_shortname - The app's short name.

$sel:dataSources:CreateApp', createApp_dataSources - The app's data source.

$sel:appSource:CreateApp', createApp_appSource - A Source object that specifies the app repository.

$sel:attributes:CreateApp', createApp_attributes - One or more user-defined key/value pairs to be added to the stack attributes.

$sel:domains:CreateApp', createApp_domains - The app virtual host settings, with multiple domains separated by commas. For example: 'www.example.com, example.com'

$sel:description:CreateApp', createApp_description - A description of the app.

$sel:stackId:CreateApp', createApp_stackId - The stack ID.

$sel:name:CreateApp', createApp_name - The app name.

$sel:type':CreateApp', createApp_type - The app type. Each supported type is associated with a particular layer. For example, PHP applications are associated with a PHP layer. AWS OpsWorks Stacks deploys an application to those instances that are members of the corresponding layer. If your app isn't one of the standard types, or you prefer to implement your own Deploy recipes, specify other.

Request Lenses

createApp_sslConfiguration :: Lens' CreateApp (Maybe SslConfiguration)

An SslConfiguration object with the SSL configuration.

createApp_environment :: Lens' CreateApp (Maybe [EnvironmentVariable])

An array of EnvironmentVariable objects that specify environment variables to be associated with the app. After you deploy the app, these variables are defined on the associated app server instance. For more information, see Environment Variables.

There is no specific limit on the number of environment variables. However, the size of the associated data structure - which includes the variables' names, values, and protected flag values - cannot exceed 20 KB. This limit should accommodate most if not all use cases. Exceeding it will cause an exception with the message, "Environment: is too large (maximum is 20KB)."

If you have specified one or more environment variables, you cannot modify the stack's Chef version.

createApp_enableSsl :: Lens' CreateApp (Maybe Bool)

Whether to enable SSL for the app.

createApp_shortname :: Lens' CreateApp (Maybe Text)

The app's short name.

createApp_dataSources :: Lens' CreateApp (Maybe [DataSource])

The app's data source.

createApp_appSource :: Lens' CreateApp (Maybe Source)

A Source object that specifies the app repository.

createApp_attributes :: Lens' CreateApp (Maybe (HashMap AppAttributesKeys Text))

One or more user-defined key/value pairs to be added to the stack attributes.

createApp_domains :: Lens' CreateApp (Maybe [Text])

The app virtual host settings, with multiple domains separated by commas. For example: 'www.example.com, example.com'

createApp_description :: Lens' CreateApp (Maybe Text)

A description of the app.

createApp_stackId :: Lens' CreateApp Text

The stack ID.

createApp_name :: Lens' CreateApp Text

The app name.

createApp_type :: Lens' CreateApp AppType

The app type. Each supported type is associated with a particular layer. For example, PHP applications are associated with a PHP layer. AWS OpsWorks Stacks deploys an application to those instances that are members of the corresponding layer. If your app isn't one of the standard types, or you prefer to implement your own Deploy recipes, specify other.

Destructuring the Response

data CreateAppResponse

Contains the response to a CreateApp request.

See: newCreateAppResponse smart constructor.

Constructors

CreateAppResponse'

Fields

appId :: Maybe Text

The app ID.

httpStatus :: Int

The response's http status code.

newCreateAppResponse

Arguments

:: Int
$sel:httpStatus:CreateAppResponse'

-> CreateAppResponse

Create a value of CreateAppResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appId:CreateAppResponse', createAppResponse_appId - The app ID.

$sel:httpStatus:CreateAppResponse', createAppResponse_httpStatus - The response's http status code.

Response Lenses

createAppResponse_appId :: Lens' CreateAppResponse (Maybe Text)

The app ID.

createAppResponse_httpStatus :: Lens' CreateAppResponse Int

The response's http status code.