Network.AWS.OpsWorks.CreateInstance

Description

Creates an instance in a specified stack. For more information, see Adding an Instance to a Layer.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Creating a Request

data CreateInstance

See: newCreateInstance smart constructor.

Constructors

CreateInstance'

Fields

installUpdatesOnBoot :: Maybe Bool

Whether to install operating system and package updates when the instance boots. The default value is true. To control when updates are installed, set this value to false. You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or by manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

We strongly recommend using the default value of true to ensure that your instances have the latest security updates.

virtualizationType :: Maybe Text

The instance's virtualization type, paravirtual or hvm.

hostname :: Maybe Text

The instance host name.

sshKeyName :: Maybe Text

The instance's Amazon EC2 key-pair name.

agentVersion :: Maybe Text

The default AWS OpsWorks Stacks agent version. You have the following options:

INHERIT - Use the stack's default agent version setting.

version_number - Use the specified agent version. This value overrides the stack's default setting. To update the agent version, edit the instance configuration and specify a new version. AWS OpsWorks Stacks then automatically installs that version on the instance.

The default setting is INHERIT. To specify an agent version, you must use the complete version number, not the abbreviated number shown on the console. For a list of available agent version numbers, call DescribeAgentVersions. AgentVersion cannot be set to Chef 12.2.

subnetId :: Maybe Text

The ID of the instance's subnet. If the stack is running in a VPC, you can use this parameter to override the stack's default subnet ID value and direct AWS OpsWorks Stacks to launch the instance in a different subnet.

ebsOptimized :: Maybe Bool

Whether to create an Amazon EBS-optimized instance.

os :: Maybe Text

The instance's operating system, which must be set to one of the following.

A supported Linux operating system: An Amazon Linux version, such as Amazon Linux 2018.03, Amazon Linux 2017.09, Amazon Linux 2017.03, Amazon Linux 2016.09, Amazon Linux 2016.03, Amazon Linux 2015.09, or Amazon Linux 2015.03.

A supported Ubuntu operating system, such as Ubuntu 16.04 LTS, Ubuntu 14.04 LTS, or Ubuntu 12.04 LTS.

CentOS Linux 7

Red Hat Enterprise Linux 7

A supported Windows operating system, such as Microsoft Windows Server 2012 R2 Base, Microsoft Windows Server 2012 R2 with SQL Server Express, Microsoft Windows Server 2012 R2 with SQL Server Standard, or Microsoft Windows Server 2012 R2 with SQL Server Web.

A custom AMI: Custom.

For more information about the supported operating systems, see AWS OpsWorks Stacks Operating Systems.

The default option is the current Amazon Linux version. If you set this parameter to Custom, you must use the CreateInstance action's AmiId parameter to specify the custom AMI that you want to use. Block device mappings are not supported if the value is Custom. For more information about supported operating systems, see Operating SystemsFor more information about how to use custom AMIs with AWS OpsWorks Stacks, see Using Custom AMIs.

availabilityZone :: Maybe Text

The instance Availability Zone. For more information, see Regions and Endpoints.

tenancy :: Maybe Text

The instance's tenancy option. The default option is no tenancy, or if the instance is running in a VPC, inherit tenancy settings from the VPC. The following are valid values for this parameter: dedicated, default, or host. Because there are costs associated with changes in tenancy options, we recommend that you research tenancy options before choosing them for your instances. For more information about dedicated hosts, see Dedicated Hosts Overview and Amazon EC2 Dedicated Hosts. For more information about dedicated instances, see Dedicated Instances and Amazon EC2 Dedicated Instances.

autoScalingType :: Maybe AutoScalingType

For load-based or time-based instances, the type. Windows stacks can use only time-based instances.

architecture :: Maybe Architecture

The instance architecture. The default option is x86_64. Instance types do not necessarily support both architectures. For a list of the architectures that are supported by the different instance types, see Instance Families and Types.

amiId :: Maybe Text

A custom AMI ID to be used to create the instance. The AMI should be based on one of the supported operating systems. For more information, see Using Custom AMIs.

If you specify a custom AMI, you must set Os to Custom.

rootDeviceType :: Maybe RootDeviceType

The instance root device type. For more information, see Storage for the Root Device.

blockDeviceMappings :: Maybe [BlockDeviceMapping]

An array of BlockDeviceMapping objects that specify the instance's block devices. For more information, see Block Device Mapping. Note that block device mappings are not supported for custom AMIs.

stackId :: Text

The stack ID.

layerIds :: [Text]

An array that contains the instance's layer IDs.

instanceType :: Text

The instance type, such as t2.micro. For a list of supported instance types, open the stack in the console, choose Instances, and choose + Instance. The Size list contains the currently supported types. For more information, see Instance Families and Types. The parameter values that you use to specify the various types are in the API Name column of the Available Instance Types table.

newCreateInstance

Arguments

:: Text
$sel:stackId:CreateInstance'

-> Text
$sel:instanceType:CreateInstance'

-> CreateInstance

Create a value of CreateInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:installUpdatesOnBoot:CreateInstance', createInstance_installUpdatesOnBoot - Whether to install operating system and package updates when the instance boots. The default value is true. To control when updates are installed, set this value to false. You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or by manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

We strongly recommend using the default value of true to ensure that your instances have the latest security updates.

$sel:virtualizationType:CreateInstance', createInstance_virtualizationType - The instance's virtualization type, paravirtual or hvm.

$sel:hostname:CreateInstance', createInstance_hostname - The instance host name.

$sel:sshKeyName:CreateInstance', createInstance_sshKeyName - The instance's Amazon EC2 key-pair name.

$sel:agentVersion:CreateInstance', createInstance_agentVersion - The default AWS OpsWorks Stacks agent version. You have the following options:

INHERIT - Use the stack's default agent version setting.

version_number - Use the specified agent version. This value overrides the stack's default setting. To update the agent version, edit the instance configuration and specify a new version. AWS OpsWorks Stacks then automatically installs that version on the instance.

The default setting is INHERIT. To specify an agent version, you must use the complete version number, not the abbreviated number shown on the console. For a list of available agent version numbers, call DescribeAgentVersions. AgentVersion cannot be set to Chef 12.2.

$sel:subnetId:CreateInstance', createInstance_subnetId - The ID of the instance's subnet. If the stack is running in a VPC, you can use this parameter to override the stack's default subnet ID value and direct AWS OpsWorks Stacks to launch the instance in a different subnet.

$sel:ebsOptimized:CreateInstance', createInstance_ebsOptimized - Whether to create an Amazon EBS-optimized instance.

$sel:os:CreateInstance', createInstance_os - The instance's operating system, which must be set to one of the following.

A supported Linux operating system: An Amazon Linux version, such as Amazon Linux 2018.03, Amazon Linux 2017.09, Amazon Linux 2017.03, Amazon Linux 2016.09, Amazon Linux 2016.03, Amazon Linux 2015.09, or Amazon Linux 2015.03.

A supported Ubuntu operating system, such as Ubuntu 16.04 LTS, Ubuntu 14.04 LTS, or Ubuntu 12.04 LTS.

CentOS Linux 7

Red Hat Enterprise Linux 7

A supported Windows operating system, such as Microsoft Windows Server 2012 R2 Base, Microsoft Windows Server 2012 R2 with SQL Server Express, Microsoft Windows Server 2012 R2 with SQL Server Standard, or Microsoft Windows Server 2012 R2 with SQL Server Web.

A custom AMI: Custom.

For more information about the supported operating systems, see AWS OpsWorks Stacks Operating Systems.

The default option is the current Amazon Linux version. If you set this parameter to Custom, you must use the CreateInstance action's AmiId parameter to specify the custom AMI that you want to use. Block device mappings are not supported if the value is Custom. For more information about supported operating systems, see Operating SystemsFor more information about how to use custom AMIs with AWS OpsWorks Stacks, see Using Custom AMIs.

$sel:availabilityZone:CreateInstance', createInstance_availabilityZone - The instance Availability Zone. For more information, see Regions and Endpoints.

$sel:tenancy:CreateInstance', createInstance_tenancy - The instance's tenancy option. The default option is no tenancy, or if the instance is running in a VPC, inherit tenancy settings from the VPC. The following are valid values for this parameter: dedicated, default, or host. Because there are costs associated with changes in tenancy options, we recommend that you research tenancy options before choosing them for your instances. For more information about dedicated hosts, see Dedicated Hosts Overview and Amazon EC2 Dedicated Hosts. For more information about dedicated instances, see Dedicated Instances and Amazon EC2 Dedicated Instances.

$sel:autoScalingType:CreateInstance', createInstance_autoScalingType - For load-based or time-based instances, the type. Windows stacks can use only time-based instances.

$sel:architecture:CreateInstance', createInstance_architecture - The instance architecture. The default option is x86_64. Instance types do not necessarily support both architectures. For a list of the architectures that are supported by the different instance types, see Instance Families and Types.

$sel:amiId:CreateInstance', createInstance_amiId - A custom AMI ID to be used to create the instance. The AMI should be based on one of the supported operating systems. For more information, see Using Custom AMIs.

If you specify a custom AMI, you must set Os to Custom.

$sel:rootDeviceType:CreateInstance', createInstance_rootDeviceType - The instance root device type. For more information, see Storage for the Root Device.

$sel:blockDeviceMappings:CreateInstance', createInstance_blockDeviceMappings - An array of BlockDeviceMapping objects that specify the instance's block devices. For more information, see Block Device Mapping. Note that block device mappings are not supported for custom AMIs.

$sel:stackId:CreateInstance', createInstance_stackId - The stack ID.

$sel:layerIds:CreateInstance', createInstance_layerIds - An array that contains the instance's layer IDs.

$sel:instanceType:CreateInstance', createInstance_instanceType - The instance type, such as t2.micro. For a list of supported instance types, open the stack in the console, choose Instances, and choose + Instance. The Size list contains the currently supported types. For more information, see Instance Families and Types. The parameter values that you use to specify the various types are in the API Name column of the Available Instance Types table.

Request Lenses

createInstance_installUpdatesOnBoot :: Lens' CreateInstance (Maybe Bool)

Whether to install operating system and package updates when the instance boots. The default value is true. To control when updates are installed, set this value to false. You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or by manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

We strongly recommend using the default value of true to ensure that your instances have the latest security updates.

createInstance_virtualizationType :: Lens' CreateInstance (Maybe Text)

The instance's virtualization type, paravirtual or hvm.

createInstance_hostname :: Lens' CreateInstance (Maybe Text)

The instance host name.

createInstance_sshKeyName :: Lens' CreateInstance (Maybe Text)

The instance's Amazon EC2 key-pair name.

createInstance_agentVersion :: Lens' CreateInstance (Maybe Text)

The default AWS OpsWorks Stacks agent version. You have the following options:

INHERIT - Use the stack's default agent version setting.

version_number - Use the specified agent version. This value overrides the stack's default setting. To update the agent version, edit the instance configuration and specify a new version. AWS OpsWorks Stacks then automatically installs that version on the instance.

The default setting is INHERIT. To specify an agent version, you must use the complete version number, not the abbreviated number shown on the console. For a list of available agent version numbers, call DescribeAgentVersions. AgentVersion cannot be set to Chef 12.2.

createInstance_subnetId :: Lens' CreateInstance (Maybe Text)

The ID of the instance's subnet. If the stack is running in a VPC, you can use this parameter to override the stack's default subnet ID value and direct AWS OpsWorks Stacks to launch the instance in a different subnet.

createInstance_ebsOptimized :: Lens' CreateInstance (Maybe Bool)

Whether to create an Amazon EBS-optimized instance.

createInstance_os :: Lens' CreateInstance (Maybe Text)

The instance's operating system, which must be set to one of the following.

A supported Linux operating system: An Amazon Linux version, such as Amazon Linux 2018.03, Amazon Linux 2017.09, Amazon Linux 2017.03, Amazon Linux 2016.09, Amazon Linux 2016.03, Amazon Linux 2015.09, or Amazon Linux 2015.03.

A supported Ubuntu operating system, such as Ubuntu 16.04 LTS, Ubuntu 14.04 LTS, or Ubuntu 12.04 LTS.

CentOS Linux 7

Red Hat Enterprise Linux 7

A supported Windows operating system, such as Microsoft Windows Server 2012 R2 Base, Microsoft Windows Server 2012 R2 with SQL Server Express, Microsoft Windows Server 2012 R2 with SQL Server Standard, or Microsoft Windows Server 2012 R2 with SQL Server Web.

A custom AMI: Custom.

For more information about the supported operating systems, see AWS OpsWorks Stacks Operating Systems.

The default option is the current Amazon Linux version. If you set this parameter to Custom, you must use the CreateInstance action's AmiId parameter to specify the custom AMI that you want to use. Block device mappings are not supported if the value is Custom. For more information about supported operating systems, see Operating SystemsFor more information about how to use custom AMIs with AWS OpsWorks Stacks, see Using Custom AMIs.

createInstance_availabilityZone :: Lens' CreateInstance (Maybe Text)

The instance Availability Zone. For more information, see Regions and Endpoints.

createInstance_tenancy :: Lens' CreateInstance (Maybe Text)

The instance's tenancy option. The default option is no tenancy, or if the instance is running in a VPC, inherit tenancy settings from the VPC. The following are valid values for this parameter: dedicated, default, or host. Because there are costs associated with changes in tenancy options, we recommend that you research tenancy options before choosing them for your instances. For more information about dedicated hosts, see Dedicated Hosts Overview and Amazon EC2 Dedicated Hosts. For more information about dedicated instances, see Dedicated Instances and Amazon EC2 Dedicated Instances.

createInstance_autoScalingType :: Lens' CreateInstance (Maybe AutoScalingType)

For load-based or time-based instances, the type. Windows stacks can use only time-based instances.

createInstance_architecture :: Lens' CreateInstance (Maybe Architecture)

The instance architecture. The default option is x86_64. Instance types do not necessarily support both architectures. For a list of the architectures that are supported by the different instance types, see Instance Families and Types.

createInstance_amiId :: Lens' CreateInstance (Maybe Text)

A custom AMI ID to be used to create the instance. The AMI should be based on one of the supported operating systems. For more information, see Using Custom AMIs.

If you specify a custom AMI, you must set Os to Custom.

createInstance_rootDeviceType :: Lens' CreateInstance (Maybe RootDeviceType)

The instance root device type. For more information, see Storage for the Root Device.

createInstance_blockDeviceMappings :: Lens' CreateInstance (Maybe [BlockDeviceMapping])

An array of BlockDeviceMapping objects that specify the instance's block devices. For more information, see Block Device Mapping. Note that block device mappings are not supported for custom AMIs.

createInstance_stackId :: Lens' CreateInstance Text

The stack ID.

createInstance_layerIds :: Lens' CreateInstance [Text]

An array that contains the instance's layer IDs.

createInstance_instanceType :: Lens' CreateInstance Text

The instance type, such as t2.micro. For a list of supported instance types, open the stack in the console, choose Instances, and choose + Instance. The Size list contains the currently supported types. For more information, see Instance Families and Types. The parameter values that you use to specify the various types are in the API Name column of the Available Instance Types table.

Destructuring the Response

data CreateInstanceResponse

Contains the response to a CreateInstance request.

See: newCreateInstanceResponse smart constructor.

Constructors

CreateInstanceResponse'

Fields

instanceId :: Maybe Text

The instance ID.

httpStatus :: Int

The response's http status code.

newCreateInstanceResponse

Arguments

:: Int
$sel:httpStatus:CreateInstanceResponse'

-> CreateInstanceResponse

Create a value of CreateInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:CreateInstanceResponse', createInstanceResponse_instanceId - The instance ID.

$sel:httpStatus:CreateInstanceResponse', createInstanceResponse_httpStatus - The response's http status code.

Response Lenses

createInstanceResponse_instanceId :: Lens' CreateInstanceResponse (Maybe Text)

The instance ID.

createInstanceResponse_httpStatus :: Lens' CreateInstanceResponse Int

The response's http status code.