Network.AWS.OpsWorks.CreateUserProfile

Description

Creates a new user profile.

Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information about user permissions, see Managing User Permissions.

Creating a Request

data CreateUserProfile

See: newCreateUserProfile smart constructor.

Constructors

CreateUserProfile'

Fields

allowSelfManagement :: Maybe Bool

Whether users can specify their own SSH public key through the My Settings page. For more information, see Setting an IAM User's Public SSH Key.

sshPublicKey :: Maybe Text

The user's public SSH key.

sshUsername :: Maybe Text

The user's SSH user name. The allowable characters are [a-z], [A-Z], [0-9], '-', and '_'. If the specified name includes other punctuation marks, AWS OpsWorks Stacks removes them. For example, my.name will be changed to myname. If you do not specify an SSH user name, AWS OpsWorks Stacks generates one from the IAM user name.

iamUserArn :: Text

The user's IAM ARN; this can also be a federated user's ARN.

newCreateUserProfile

Arguments

:: Text
$sel:iamUserArn:CreateUserProfile'

-> CreateUserProfile

Create a value of CreateUserProfile with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:allowSelfManagement:CreateUserProfile', createUserProfile_allowSelfManagement - Whether users can specify their own SSH public key through the My Settings page. For more information, see Setting an IAM User's Public SSH Key.

$sel:sshPublicKey:CreateUserProfile', createUserProfile_sshPublicKey - The user's public SSH key.

$sel:sshUsername:CreateUserProfile', createUserProfile_sshUsername - The user's SSH user name. The allowable characters are [a-z], [A-Z], [0-9], '-', and '_'. If the specified name includes other punctuation marks, AWS OpsWorks Stacks removes them. For example, my.name will be changed to myname. If you do not specify an SSH user name, AWS OpsWorks Stacks generates one from the IAM user name.

$sel:iamUserArn:CreateUserProfile', createUserProfile_iamUserArn - The user's IAM ARN; this can also be a federated user's ARN.

Request Lenses

createUserProfile_allowSelfManagement :: Lens' CreateUserProfile (Maybe Bool)

Whether users can specify their own SSH public key through the My Settings page. For more information, see Setting an IAM User's Public SSH Key.

createUserProfile_sshPublicKey :: Lens' CreateUserProfile (Maybe Text)

The user's public SSH key.

createUserProfile_sshUsername :: Lens' CreateUserProfile (Maybe Text)

The user's SSH user name. The allowable characters are [a-z], [A-Z], [0-9], '-', and '_'. If the specified name includes other punctuation marks, AWS OpsWorks Stacks removes them. For example, my.name will be changed to myname. If you do not specify an SSH user name, AWS OpsWorks Stacks generates one from the IAM user name.

createUserProfile_iamUserArn :: Lens' CreateUserProfile Text

The user's IAM ARN; this can also be a federated user's ARN.

Destructuring the Response

data CreateUserProfileResponse

Contains the response to a CreateUserProfile request.

See: newCreateUserProfileResponse smart constructor.

Constructors

CreateUserProfileResponse'

Fields

iamUserArn :: Maybe Text

The user's IAM ARN.

httpStatus :: Int

The response's http status code.

newCreateUserProfileResponse

Arguments

:: Int
$sel:httpStatus:CreateUserProfileResponse'

-> CreateUserProfileResponse

Create a value of CreateUserProfileResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iamUserArn:CreateUserProfile', createUserProfileResponse_iamUserArn - The user's IAM ARN.

$sel:httpStatus:CreateUserProfileResponse', createUserProfileResponse_httpStatus - The response's http status code.

Response Lenses

createUserProfileResponse_iamUserArn :: Lens' CreateUserProfileResponse (Maybe Text)

The user's IAM ARN.

createUserProfileResponse_httpStatus :: Lens' CreateUserProfileResponse Int

The response's http status code.