Network.AWS.OpsWorks.DeregisterElasticIp

Description

Deregisters a specified Elastic IP address. The address can then be registered by another stack. For more information, see Resource Management.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Creating a Request

data DeregisterElasticIp

See: newDeregisterElasticIp smart constructor.

Constructors

DeregisterElasticIp'

Fields

elasticIp :: Text

The Elastic IP address.

newDeregisterElasticIp

Arguments

:: Text
$sel:elasticIp:DeregisterElasticIp'

-> DeregisterElasticIp

Create a value of DeregisterElasticIp with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:elasticIp:DeregisterElasticIp', deregisterElasticIp_elasticIp - The Elastic IP address.

Request Lenses

deregisterElasticIp_elasticIp :: Lens' DeregisterElasticIp Text

The Elastic IP address.

Destructuring the Response

data DeregisterElasticIpResponse

See: newDeregisterElasticIpResponse smart constructor.

Constructors

DeregisterElasticIpResponse'

newDeregisterElasticIpResponse :: DeregisterElasticIpResponse

Create a value of DeregisterElasticIpResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.