Network.AWS.OpsWorks.DescribeApps

Description

Requests a description of a specified set of apps.

This call accepts only one resource-identifying parameter.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information about user permissions, see Managing User Permissions.

Creating a Request

data DescribeApps

See: newDescribeApps smart constructor.

Constructors

DescribeApps'

Fields

appIds :: Maybe [Text]

An array of app IDs for the apps to be described. If you use this parameter, DescribeApps returns a description of the specified apps. Otherwise, it returns a description of every app.

stackId :: Maybe Text

The app stack ID. If you use this parameter, DescribeApps returns a description of the apps in the specified stack.

newDescribeApps :: DescribeApps

Create a value of DescribeApps with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:appIds:DescribeApps', describeApps_appIds - An array of app IDs for the apps to be described. If you use this parameter, DescribeApps returns a description of the specified apps. Otherwise, it returns a description of every app.

$sel:stackId:DescribeApps', describeApps_stackId - The app stack ID. If you use this parameter, DescribeApps returns a description of the apps in the specified stack.

Request Lenses

describeApps_appIds :: Lens' DescribeApps (Maybe [Text])

An array of app IDs for the apps to be described. If you use this parameter, DescribeApps returns a description of the specified apps. Otherwise, it returns a description of every app.

describeApps_stackId :: Lens' DescribeApps (Maybe Text)

The app stack ID. If you use this parameter, DescribeApps returns a description of the apps in the specified stack.

Destructuring the Response

data DescribeAppsResponse

Contains the response to a DescribeApps request.

See: newDescribeAppsResponse smart constructor.

Constructors

DescribeAppsResponse'

Fields

apps :: Maybe [App]

An array of App objects that describe the specified apps.

httpStatus :: Int

The response's http status code.

newDescribeAppsResponse

Arguments

:: Int
$sel:httpStatus:DescribeAppsResponse'

-> DescribeAppsResponse

Create a value of DescribeAppsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:apps:DescribeAppsResponse', describeAppsResponse_apps - An array of App objects that describe the specified apps.

$sel:httpStatus:DescribeAppsResponse', describeAppsResponse_httpStatus - The response's http status code.

Response Lenses

describeAppsResponse_apps :: Lens' DescribeAppsResponse (Maybe [App])

An array of App objects that describe the specified apps.

describeAppsResponse_httpStatus :: Lens' DescribeAppsResponse Int

The response's http status code.