Network.AWS.OpsWorks.DescribeCommands

Description

Describes the results of specified commands.

This call accepts only one resource-identifying parameter.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information about user permissions, see Managing User Permissions.

Creating a Request

data DescribeCommands

See: newDescribeCommands smart constructor.

Constructors

DescribeCommands'

Fields

deploymentId :: Maybe Text

The deployment ID. If you include this parameter, DescribeCommands returns a description of the commands associated with the specified deployment.

instanceId :: Maybe Text

The instance ID. If you include this parameter, DescribeCommands returns a description of the commands associated with the specified instance.

commandIds :: Maybe [Text]

An array of command IDs. If you include this parameter, DescribeCommands returns a description of the specified commands. Otherwise, it returns a description of every command.

newDescribeCommands :: DescribeCommands

Create a value of DescribeCommands with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentId:DescribeCommands', describeCommands_deploymentId - The deployment ID. If you include this parameter, DescribeCommands returns a description of the commands associated with the specified deployment.

$sel:instanceId:DescribeCommands', describeCommands_instanceId - The instance ID. If you include this parameter, DescribeCommands returns a description of the commands associated with the specified instance.

$sel:commandIds:DescribeCommands', describeCommands_commandIds - An array of command IDs. If you include this parameter, DescribeCommands returns a description of the specified commands. Otherwise, it returns a description of every command.

Request Lenses

describeCommands_deploymentId :: Lens' DescribeCommands (Maybe Text)

The deployment ID. If you include this parameter, DescribeCommands returns a description of the commands associated with the specified deployment.

describeCommands_instanceId :: Lens' DescribeCommands (Maybe Text)

The instance ID. If you include this parameter, DescribeCommands returns a description of the commands associated with the specified instance.

describeCommands_commandIds :: Lens' DescribeCommands (Maybe [Text])

An array of command IDs. If you include this parameter, DescribeCommands returns a description of the specified commands. Otherwise, it returns a description of every command.

Destructuring the Response

data DescribeCommandsResponse

Contains the response to a DescribeCommands request.

See: newDescribeCommandsResponse smart constructor.

Constructors

DescribeCommandsResponse'

Fields

commands :: Maybe [Command]

An array of Command objects that describe each of the specified commands.

httpStatus :: Int

The response's http status code.

newDescribeCommandsResponse

Arguments

:: Int
$sel:httpStatus:DescribeCommandsResponse'

-> DescribeCommandsResponse

Create a value of DescribeCommandsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:commands:DescribeCommandsResponse', describeCommandsResponse_commands - An array of Command objects that describe each of the specified commands.

$sel:httpStatus:DescribeCommandsResponse', describeCommandsResponse_httpStatus - The response's http status code.

Response Lenses

describeCommandsResponse_commands :: Lens' DescribeCommandsResponse (Maybe [Command])

An array of Command objects that describe each of the specified commands.

describeCommandsResponse_httpStatus :: Lens' DescribeCommandsResponse Int

The response's http status code.