Network.AWS.OpsWorks.DescribeElasticIps

Description

Describes Elastic IP addresses.

This call accepts only one resource-identifying parameter.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information about user permissions, see Managing User Permissions.

Creating a Request

data DescribeElasticIps

See: newDescribeElasticIps smart constructor.

Constructors

DescribeElasticIps'

Fields

instanceId :: Maybe Text

The instance ID. If you include this parameter, DescribeElasticIps returns a description of the Elastic IP addresses associated with the specified instance.

ips :: Maybe [Text]

An array of Elastic IP addresses to be described. If you include this parameter, DescribeElasticIps returns a description of the specified Elastic IP addresses. Otherwise, it returns a description of every Elastic IP address.

stackId :: Maybe Text

A stack ID. If you include this parameter, DescribeElasticIps returns a description of the Elastic IP addresses that are registered with the specified stack.

newDescribeElasticIps :: DescribeElasticIps

Create a value of DescribeElasticIps with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:DescribeElasticIps', describeElasticIps_instanceId - The instance ID. If you include this parameter, DescribeElasticIps returns a description of the Elastic IP addresses associated with the specified instance.

$sel:ips:DescribeElasticIps', describeElasticIps_ips - An array of Elastic IP addresses to be described. If you include this parameter, DescribeElasticIps returns a description of the specified Elastic IP addresses. Otherwise, it returns a description of every Elastic IP address.

$sel:stackId:DescribeElasticIps', describeElasticIps_stackId - A stack ID. If you include this parameter, DescribeElasticIps returns a description of the Elastic IP addresses that are registered with the specified stack.

Request Lenses

describeElasticIps_instanceId :: Lens' DescribeElasticIps (Maybe Text)

The instance ID. If you include this parameter, DescribeElasticIps returns a description of the Elastic IP addresses associated with the specified instance.

describeElasticIps_ips :: Lens' DescribeElasticIps (Maybe [Text])

An array of Elastic IP addresses to be described. If you include this parameter, DescribeElasticIps returns a description of the specified Elastic IP addresses. Otherwise, it returns a description of every Elastic IP address.

describeElasticIps_stackId :: Lens' DescribeElasticIps (Maybe Text)

A stack ID. If you include this parameter, DescribeElasticIps returns a description of the Elastic IP addresses that are registered with the specified stack.

Destructuring the Response

data DescribeElasticIpsResponse

Contains the response to a DescribeElasticIps request.

See: newDescribeElasticIpsResponse smart constructor.

Constructors

DescribeElasticIpsResponse'

Fields

elasticIps :: Maybe [ElasticIp]

An ElasticIps object that describes the specified Elastic IP addresses.

httpStatus :: Int

The response's http status code.

newDescribeElasticIpsResponse

Arguments

:: Int
$sel:httpStatus:DescribeElasticIpsResponse'

-> DescribeElasticIpsResponse

Create a value of DescribeElasticIpsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:elasticIps:DescribeElasticIpsResponse', describeElasticIpsResponse_elasticIps - An ElasticIps object that describes the specified Elastic IP addresses.

$sel:httpStatus:DescribeElasticIpsResponse', describeElasticIpsResponse_httpStatus - The response's http status code.

Response Lenses

describeElasticIpsResponse_elasticIps :: Lens' DescribeElasticIpsResponse (Maybe [ElasticIp])

An ElasticIps object that describes the specified Elastic IP addresses.

describeElasticIpsResponse_httpStatus :: Lens' DescribeElasticIpsResponse Int

The response's http status code.