Network.AWS.OpsWorks.DescribeElasticLoadBalancers

Description

Describes a stack's Elastic Load Balancing instances.

This call accepts only one resource-identifying parameter.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information about user permissions, see Managing User Permissions.

Creating a Request

data DescribeElasticLoadBalancers

See: newDescribeElasticLoadBalancers smart constructor.

Constructors

DescribeElasticLoadBalancers'

Fields

layerIds :: Maybe [Text]

A list of layer IDs. The action describes the Elastic Load Balancing instances for the specified layers.

stackId :: Maybe Text

A stack ID. The action describes the stack's Elastic Load Balancing instances.

newDescribeElasticLoadBalancers :: DescribeElasticLoadBalancers

Create a value of DescribeElasticLoadBalancers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:layerIds:DescribeElasticLoadBalancers', describeElasticLoadBalancers_layerIds - A list of layer IDs. The action describes the Elastic Load Balancing instances for the specified layers.

$sel:stackId:DescribeElasticLoadBalancers', describeElasticLoadBalancers_stackId - A stack ID. The action describes the stack's Elastic Load Balancing instances.

Request Lenses

describeElasticLoadBalancers_layerIds :: Lens' DescribeElasticLoadBalancers (Maybe [Text])

A list of layer IDs. The action describes the Elastic Load Balancing instances for the specified layers.

describeElasticLoadBalancers_stackId :: Lens' DescribeElasticLoadBalancers (Maybe Text)

A stack ID. The action describes the stack's Elastic Load Balancing instances.

Destructuring the Response

data DescribeElasticLoadBalancersResponse

Contains the response to a DescribeElasticLoadBalancers request.

See: newDescribeElasticLoadBalancersResponse smart constructor.

Constructors

DescribeElasticLoadBalancersResponse'

Fields

elasticLoadBalancers :: Maybe [ElasticLoadBalancer]

A list of ElasticLoadBalancer objects that describe the specified Elastic Load Balancing instances.

httpStatus :: Int

The response's http status code.

newDescribeElasticLoadBalancersResponse

Arguments

:: Int
$sel:httpStatus:DescribeElasticLoadBalancersResponse'

-> DescribeElasticLoadBalancersResponse

Create a value of DescribeElasticLoadBalancersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:elasticLoadBalancers:DescribeElasticLoadBalancersResponse', describeElasticLoadBalancersResponse_elasticLoadBalancers - A list of ElasticLoadBalancer objects that describe the specified Elastic Load Balancing instances.

$sel:httpStatus:DescribeElasticLoadBalancersResponse', describeElasticLoadBalancersResponse_httpStatus - The response's http status code.

Response Lenses

describeElasticLoadBalancersResponse_elasticLoadBalancers :: Lens' DescribeElasticLoadBalancersResponse (Maybe [ElasticLoadBalancer])

A list of ElasticLoadBalancer objects that describe the specified Elastic Load Balancing instances.

describeElasticLoadBalancersResponse_httpStatus :: Lens' DescribeElasticLoadBalancersResponse Int

The response's http status code.