Network.AWS.OpsWorks.DescribeLoadBasedAutoScaling

Description

Describes load-based auto scaling configurations for specified layers.

You must specify at least one of the parameters.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information about user permissions, see Managing User Permissions.

Creating a Request

data DescribeLoadBasedAutoScaling

See: newDescribeLoadBasedAutoScaling smart constructor.

Constructors

DescribeLoadBasedAutoScaling'

Fields

layerIds :: [Text]

An array of layer IDs.

newDescribeLoadBasedAutoScaling :: DescribeLoadBasedAutoScaling

Create a value of DescribeLoadBasedAutoScaling with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:layerIds:DescribeLoadBasedAutoScaling', describeLoadBasedAutoScaling_layerIds - An array of layer IDs.

Request Lenses

describeLoadBasedAutoScaling_layerIds :: Lens' DescribeLoadBasedAutoScaling [Text]

An array of layer IDs.

Destructuring the Response

data DescribeLoadBasedAutoScalingResponse

Contains the response to a DescribeLoadBasedAutoScaling request.

See: newDescribeLoadBasedAutoScalingResponse smart constructor.

Constructors

DescribeLoadBasedAutoScalingResponse'

Fields

loadBasedAutoScalingConfigurations :: Maybe [LoadBasedAutoScalingConfiguration]

An array of LoadBasedAutoScalingConfiguration objects that describe each layer's configuration.

httpStatus :: Int

The response's http status code.

newDescribeLoadBasedAutoScalingResponse

Arguments

:: Int
$sel:httpStatus:DescribeLoadBasedAutoScalingResponse'

-> DescribeLoadBasedAutoScalingResponse

Create a value of DescribeLoadBasedAutoScalingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:loadBasedAutoScalingConfigurations:DescribeLoadBasedAutoScalingResponse', describeLoadBasedAutoScalingResponse_loadBasedAutoScalingConfigurations - An array of LoadBasedAutoScalingConfiguration objects that describe each layer's configuration.

$sel:httpStatus:DescribeLoadBasedAutoScalingResponse', describeLoadBasedAutoScalingResponse_httpStatus - The response's http status code.

Response Lenses

describeLoadBasedAutoScalingResponse_loadBasedAutoScalingConfigurations :: Lens' DescribeLoadBasedAutoScalingResponse (Maybe [LoadBasedAutoScalingConfiguration])

An array of LoadBasedAutoScalingConfiguration objects that describe each layer's configuration.

describeLoadBasedAutoScalingResponse_httpStatus :: Lens' DescribeLoadBasedAutoScalingResponse Int

The response's http status code.