Network.AWS.OpsWorks.DescribeOperatingSystems

Description

Describes the operating systems that are supported by AWS OpsWorks Stacks.

Creating a Request

data DescribeOperatingSystems

See: newDescribeOperatingSystems smart constructor.

Constructors

DescribeOperatingSystems'

newDescribeOperatingSystems :: DescribeOperatingSystems

Create a value of DescribeOperatingSystems with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data DescribeOperatingSystemsResponse

The response to a DescribeOperatingSystems request.

See: newDescribeOperatingSystemsResponse smart constructor.

Constructors

DescribeOperatingSystemsResponse'

Fields

operatingSystems :: Maybe [OperatingSystem]

Contains information in response to a DescribeOperatingSystems request.

httpStatus :: Int

The response's http status code.

newDescribeOperatingSystemsResponse

Arguments

:: Int
$sel:httpStatus:DescribeOperatingSystemsResponse'

-> DescribeOperatingSystemsResponse

Create a value of DescribeOperatingSystemsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operatingSystems:DescribeOperatingSystemsResponse', describeOperatingSystemsResponse_operatingSystems - Contains information in response to a DescribeOperatingSystems request.

$sel:httpStatus:DescribeOperatingSystemsResponse', describeOperatingSystemsResponse_httpStatus - The response's http status code.

Response Lenses

describeOperatingSystemsResponse_operatingSystems :: Lens' DescribeOperatingSystemsResponse (Maybe [OperatingSystem])

Contains information in response to a DescribeOperatingSystems request.

describeOperatingSystemsResponse_httpStatus :: Lens' DescribeOperatingSystemsResponse Int

The response's http status code.