Network.AWS.OpsWorks.DescribePermissions

Description

Describes the permissions for a specified stack.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Creating a Request

data DescribePermissions

See: newDescribePermissions smart constructor.

Constructors

DescribePermissions'

Fields

iamUserArn :: Maybe Text

The user's IAM ARN. This can also be a federated user's ARN. For more information about IAM ARNs, see Using Identifiers.

stackId :: Maybe Text

The stack ID.

newDescribePermissions :: DescribePermissions

Create a value of DescribePermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iamUserArn:DescribePermissions', describePermissions_iamUserArn - The user's IAM ARN. This can also be a federated user's ARN. For more information about IAM ARNs, see Using Identifiers.

$sel:stackId:DescribePermissions', describePermissions_stackId - The stack ID.

Request Lenses

describePermissions_iamUserArn :: Lens' DescribePermissions (Maybe Text)

The user's IAM ARN. This can also be a federated user's ARN. For more information about IAM ARNs, see Using Identifiers.

describePermissions_stackId :: Lens' DescribePermissions (Maybe Text)

The stack ID.

Destructuring the Response

data DescribePermissionsResponse

Contains the response to a DescribePermissions request.

See: newDescribePermissionsResponse smart constructor.

Constructors

DescribePermissionsResponse'

Fields

permissions :: Maybe [Permission]

An array of Permission objects that describe the stack permissions.

If the request object contains only a stack ID, the array contains a Permission object with permissions for each of the stack IAM ARNs.

If the request object contains only an IAM ARN, the array contains a Permission object with permissions for each of the user's stack IDs.

If the request contains a stack ID and an IAM ARN, the array contains a single Permission object with permissions for the specified stack and IAM ARN.

httpStatus :: Int

The response's http status code.

newDescribePermissionsResponse

Arguments

:: Int
$sel:httpStatus:DescribePermissionsResponse'

-> DescribePermissionsResponse

Create a value of DescribePermissionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:permissions:DescribePermissionsResponse', describePermissionsResponse_permissions - An array of Permission objects that describe the stack permissions.

If the request object contains only a stack ID, the array contains a Permission object with permissions for each of the stack IAM ARNs.

If the request object contains only an IAM ARN, the array contains a Permission object with permissions for each of the user's stack IDs.

If the request contains a stack ID and an IAM ARN, the array contains a single Permission object with permissions for the specified stack and IAM ARN.

$sel:httpStatus:DescribePermissionsResponse', describePermissionsResponse_httpStatus - The response's http status code.

Response Lenses

describePermissionsResponse_permissions :: Lens' DescribePermissionsResponse (Maybe [Permission])

An array of Permission objects that describe the stack permissions.

If the request object contains only a stack ID, the array contains a Permission object with permissions for each of the stack IAM ARNs.

If the request object contains only an IAM ARN, the array contains a Permission object with permissions for each of the user's stack IDs.

If the request contains a stack ID and an IAM ARN, the array contains a single Permission object with permissions for the specified stack and IAM ARN.

describePermissionsResponse_httpStatus :: Lens' DescribePermissionsResponse Int

The response's http status code.