Network.AWS.OpsWorks.DescribeRaidArrays

Description

Describe an instance's RAID arrays.

This call accepts only one resource-identifying parameter.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information about user permissions, see Managing User Permissions.

Creating a Request

data DescribeRaidArrays

See: newDescribeRaidArrays smart constructor.

Constructors

DescribeRaidArrays'

Fields

instanceId :: Maybe Text

The instance ID. If you use this parameter, DescribeRaidArrays returns descriptions of the RAID arrays associated with the specified instance.

raidArrayIds :: Maybe [Text]

An array of RAID array IDs. If you use this parameter, DescribeRaidArrays returns descriptions of the specified arrays. Otherwise, it returns a description of every array.

stackId :: Maybe Text

The stack ID.

newDescribeRaidArrays :: DescribeRaidArrays

Create a value of DescribeRaidArrays with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:DescribeRaidArrays', describeRaidArrays_instanceId - The instance ID. If you use this parameter, DescribeRaidArrays returns descriptions of the RAID arrays associated with the specified instance.

$sel:raidArrayIds:DescribeRaidArrays', describeRaidArrays_raidArrayIds - An array of RAID array IDs. If you use this parameter, DescribeRaidArrays returns descriptions of the specified arrays. Otherwise, it returns a description of every array.

$sel:stackId:DescribeRaidArrays', describeRaidArrays_stackId - The stack ID.

Request Lenses

describeRaidArrays_instanceId :: Lens' DescribeRaidArrays (Maybe Text)

The instance ID. If you use this parameter, DescribeRaidArrays returns descriptions of the RAID arrays associated with the specified instance.

describeRaidArrays_raidArrayIds :: Lens' DescribeRaidArrays (Maybe [Text])

An array of RAID array IDs. If you use this parameter, DescribeRaidArrays returns descriptions of the specified arrays. Otherwise, it returns a description of every array.

describeRaidArrays_stackId :: Lens' DescribeRaidArrays (Maybe Text)

The stack ID.

Destructuring the Response

data DescribeRaidArraysResponse

Contains the response to a DescribeRaidArrays request.

See: newDescribeRaidArraysResponse smart constructor.

Constructors

DescribeRaidArraysResponse'

Fields

raidArrays :: Maybe [RaidArray]

A RaidArrays object that describes the specified RAID arrays.

httpStatus :: Int

The response's http status code.

newDescribeRaidArraysResponse

Arguments

:: Int
$sel:httpStatus:DescribeRaidArraysResponse'

-> DescribeRaidArraysResponse

Create a value of DescribeRaidArraysResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:raidArrays:DescribeRaidArraysResponse', describeRaidArraysResponse_raidArrays - A RaidArrays object that describes the specified RAID arrays.

$sel:httpStatus:DescribeRaidArraysResponse', describeRaidArraysResponse_httpStatus - The response's http status code.

Response Lenses

describeRaidArraysResponse_raidArrays :: Lens' DescribeRaidArraysResponse (Maybe [RaidArray])

A RaidArrays object that describes the specified RAID arrays.

describeRaidArraysResponse_httpStatus :: Lens' DescribeRaidArraysResponse Int

The response's http status code.