Network.AWS.OpsWorks.DescribeRdsDbInstances

Description

Describes Amazon RDS instances.

Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information about user permissions, see Managing User Permissions.

This call accepts only one resource-identifying parameter.

Creating a Request

data DescribeRdsDbInstances

See: newDescribeRdsDbInstances smart constructor.

Constructors

DescribeRdsDbInstances'

Fields

rdsDbInstanceArns :: Maybe [Text]

An array containing the ARNs of the instances to be described.

stackId :: Text

The ID of the stack with which the instances are registered. The operation returns descriptions of all registered Amazon RDS instances.

newDescribeRdsDbInstances

Arguments

:: Text
$sel:stackId:DescribeRdsDbInstances'

-> DescribeRdsDbInstances

Create a value of DescribeRdsDbInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rdsDbInstanceArns:DescribeRdsDbInstances', describeRdsDbInstances_rdsDbInstanceArns - An array containing the ARNs of the instances to be described.

$sel:stackId:DescribeRdsDbInstances', describeRdsDbInstances_stackId - The ID of the stack with which the instances are registered. The operation returns descriptions of all registered Amazon RDS instances.

Request Lenses

describeRdsDbInstances_rdsDbInstanceArns :: Lens' DescribeRdsDbInstances (Maybe [Text])

An array containing the ARNs of the instances to be described.

describeRdsDbInstances_stackId :: Lens' DescribeRdsDbInstances Text

The ID of the stack with which the instances are registered. The operation returns descriptions of all registered Amazon RDS instances.

Destructuring the Response

data DescribeRdsDbInstancesResponse

Contains the response to a DescribeRdsDbInstances request.

See: newDescribeRdsDbInstancesResponse smart constructor.

Constructors

DescribeRdsDbInstancesResponse'

Fields

rdsDbInstances :: Maybe [RdsDbInstance]

An a array of RdsDbInstance objects that describe the instances.

httpStatus :: Int

The response's http status code.

newDescribeRdsDbInstancesResponse

Arguments

:: Int
$sel:httpStatus:DescribeRdsDbInstancesResponse'

-> DescribeRdsDbInstancesResponse

Create a value of DescribeRdsDbInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rdsDbInstances:DescribeRdsDbInstancesResponse', describeRdsDbInstancesResponse_rdsDbInstances - An a array of RdsDbInstance objects that describe the instances.

$sel:httpStatus:DescribeRdsDbInstancesResponse', describeRdsDbInstancesResponse_httpStatus - The response's http status code.

Response Lenses

describeRdsDbInstancesResponse_rdsDbInstances :: Lens' DescribeRdsDbInstancesResponse (Maybe [RdsDbInstance])

An a array of RdsDbInstance objects that describe the instances.

describeRdsDbInstancesResponse_httpStatus :: Lens' DescribeRdsDbInstancesResponse Int

The response's http status code.