Network.AWS.OpsWorks.GrantAccess

Description

This action can be used only with Windows stacks.

Grants RDP access to a Windows instance for a specified time period.

Creating a Request

data GrantAccess

See: newGrantAccess smart constructor.

Constructors

GrantAccess'

Fields

validForInMinutes :: Maybe Natural

The length of time (in minutes) that the grant is valid. When the grant expires at the end of this period, the user will no longer be able to use the credentials to log in. If the user is logged in at the time, he or she automatically will be logged out.

instanceId :: Text

The instance's AWS OpsWorks Stacks ID.

newGrantAccess

Arguments

:: Text
$sel:instanceId:GrantAccess'

-> GrantAccess

Create a value of GrantAccess with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:validForInMinutes:GrantAccess', grantAccess_validForInMinutes - The length of time (in minutes) that the grant is valid. When the grant expires at the end of this period, the user will no longer be able to use the credentials to log in. If the user is logged in at the time, he or she automatically will be logged out.

$sel:instanceId:GrantAccess', grantAccess_instanceId - The instance's AWS OpsWorks Stacks ID.

Request Lenses

grantAccess_validForInMinutes :: Lens' GrantAccess (Maybe Natural)

The length of time (in minutes) that the grant is valid. When the grant expires at the end of this period, the user will no longer be able to use the credentials to log in. If the user is logged in at the time, he or she automatically will be logged out.

grantAccess_instanceId :: Lens' GrantAccess Text

The instance's AWS OpsWorks Stacks ID.

Destructuring the Response

data GrantAccessResponse

Contains the response to a GrantAccess request.

See: newGrantAccessResponse smart constructor.

Constructors

GrantAccessResponse'

Fields

temporaryCredential :: Maybe TemporaryCredential

A TemporaryCredential object that contains the data needed to log in to the instance by RDP clients, such as the Microsoft Remote Desktop Connection.

httpStatus :: Int

The response's http status code.

newGrantAccessResponse

Arguments

:: Int
$sel:httpStatus:GrantAccessResponse'

-> GrantAccessResponse

Create a value of GrantAccessResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:temporaryCredential:GrantAccessResponse', grantAccessResponse_temporaryCredential - A TemporaryCredential object that contains the data needed to log in to the instance by RDP clients, such as the Microsoft Remote Desktop Connection.

$sel:httpStatus:GrantAccessResponse', grantAccessResponse_httpStatus - The response's http status code.

Response Lenses

grantAccessResponse_temporaryCredential :: Lens' GrantAccessResponse (Maybe TemporaryCredential)

A TemporaryCredential object that contains the data needed to log in to the instance by RDP clients, such as the Microsoft Remote Desktop Connection.

grantAccessResponse_httpStatus :: Lens' GrantAccessResponse Int

The response's http status code.