Network.AWS.OpsWorks.RegisterElasticIp

Description

Registers an Elastic IP address with a specified stack. An address can be registered with only one stack at a time. If the address is already registered, you must first deregister it by calling DeregisterElasticIp. For more information, see Resource Management.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Creating a Request

data RegisterElasticIp

See: newRegisterElasticIp smart constructor.

Constructors

RegisterElasticIp'

Fields

elasticIp :: Text

The Elastic IP address.

stackId :: Text

The stack ID.

newRegisterElasticIp

Arguments

:: Text
$sel:elasticIp:RegisterElasticIp'

-> Text
$sel:stackId:RegisterElasticIp'

-> RegisterElasticIp

Create a value of RegisterElasticIp with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:elasticIp:RegisterElasticIp', registerElasticIp_elasticIp - The Elastic IP address.

$sel:stackId:RegisterElasticIp', registerElasticIp_stackId - The stack ID.

Request Lenses

registerElasticIp_elasticIp :: Lens' RegisterElasticIp Text

The Elastic IP address.

registerElasticIp_stackId :: Lens' RegisterElasticIp Text

The stack ID.

Destructuring the Response

data RegisterElasticIpResponse

Contains the response to a RegisterElasticIp request.

See: newRegisterElasticIpResponse smart constructor.

Constructors

RegisterElasticIpResponse'

Fields

elasticIp :: Maybe Text

The Elastic IP address.

httpStatus :: Int

The response's http status code.

newRegisterElasticIpResponse

Arguments

:: Int
$sel:httpStatus:RegisterElasticIpResponse'

-> RegisterElasticIpResponse

Create a value of RegisterElasticIpResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:elasticIp:RegisterElasticIp', registerElasticIpResponse_elasticIp - The Elastic IP address.

$sel:httpStatus:RegisterElasticIpResponse', registerElasticIpResponse_httpStatus - The response's http status code.

Response Lenses

registerElasticIpResponse_elasticIp :: Lens' RegisterElasticIpResponse (Maybe Text)

The Elastic IP address.

registerElasticIpResponse_httpStatus :: Lens' RegisterElasticIpResponse Int

The response's http status code.