Network.AWS.OpsWorks.SetLoadBasedAutoScaling

Description

Specify the load-based auto scaling configuration for a specified layer. For more information, see Managing Load with Time-based and Load-based Instances.

To use load-based auto scaling, you must create a set of load-based auto scaling instances. Load-based auto scaling operates only on the instances from that set, so you must ensure that you have created enough instances to handle the maximum anticipated load.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Creating a Request

data SetLoadBasedAutoScaling

See: newSetLoadBasedAutoScaling smart constructor.

Constructors

SetLoadBasedAutoScaling'

Fields

upScaling :: Maybe AutoScalingThresholds

An AutoScalingThresholds object with the upscaling threshold configuration. If the load exceeds these thresholds for a specified amount of time, AWS OpsWorks Stacks starts a specified number of instances.

enable :: Maybe Bool

Enables load-based auto scaling for the layer.

downScaling :: Maybe AutoScalingThresholds

An AutoScalingThresholds object with the downscaling threshold configuration. If the load falls below these thresholds for a specified amount of time, AWS OpsWorks Stacks stops a specified number of instances.

layerId :: Text

The layer ID.

newSetLoadBasedAutoScaling

Arguments

:: Text
$sel:layerId:SetLoadBasedAutoScaling'

-> SetLoadBasedAutoScaling

Create a value of SetLoadBasedAutoScaling with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:upScaling:SetLoadBasedAutoScaling', setLoadBasedAutoScaling_upScaling - An AutoScalingThresholds object with the upscaling threshold configuration. If the load exceeds these thresholds for a specified amount of time, AWS OpsWorks Stacks starts a specified number of instances.

$sel:enable:SetLoadBasedAutoScaling', setLoadBasedAutoScaling_enable - Enables load-based auto scaling for the layer.

$sel:downScaling:SetLoadBasedAutoScaling', setLoadBasedAutoScaling_downScaling - An AutoScalingThresholds object with the downscaling threshold configuration. If the load falls below these thresholds for a specified amount of time, AWS OpsWorks Stacks stops a specified number of instances.

$sel:layerId:SetLoadBasedAutoScaling', setLoadBasedAutoScaling_layerId - The layer ID.

Request Lenses

setLoadBasedAutoScaling_upScaling :: Lens' SetLoadBasedAutoScaling (Maybe AutoScalingThresholds)

An AutoScalingThresholds object with the upscaling threshold configuration. If the load exceeds these thresholds for a specified amount of time, AWS OpsWorks Stacks starts a specified number of instances.

setLoadBasedAutoScaling_enable :: Lens' SetLoadBasedAutoScaling (Maybe Bool)

Enables load-based auto scaling for the layer.

setLoadBasedAutoScaling_downScaling :: Lens' SetLoadBasedAutoScaling (Maybe AutoScalingThresholds)

An AutoScalingThresholds object with the downscaling threshold configuration. If the load falls below these thresholds for a specified amount of time, AWS OpsWorks Stacks stops a specified number of instances.

setLoadBasedAutoScaling_layerId :: Lens' SetLoadBasedAutoScaling Text

The layer ID.

Destructuring the Response

data SetLoadBasedAutoScalingResponse

See: newSetLoadBasedAutoScalingResponse smart constructor.

Constructors

SetLoadBasedAutoScalingResponse'

newSetLoadBasedAutoScalingResponse :: SetLoadBasedAutoScalingResponse

Create a value of SetLoadBasedAutoScalingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.