Network.AWS.OpsWorks.SetPermission

Description

Specifies a user's permissions. For more information, see Security and Permissions.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Creating a Request

data SetPermission

See: newSetPermission smart constructor.

Constructors

SetPermission'

Fields

allowSudo :: Maybe Bool

The user is allowed to use sudo to elevate privileges.

level :: Maybe Text

The user's permission level, which must be set to one of the following strings. You cannot set your own permissions level.

deny

show

deploy

manage

iam_only

For more information about the permissions associated with these levels, see Managing User Permissions.

allowSsh :: Maybe Bool

The user is allowed to use SSH to communicate with the instance.

stackId :: Text

The stack ID.

iamUserArn :: Text

The user's IAM ARN. This can also be a federated user's ARN.

newSetPermission

Arguments

:: Text
$sel:stackId:SetPermission'

-> Text
$sel:iamUserArn:SetPermission'

-> SetPermission

Create a value of SetPermission with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:allowSudo:SetPermission', setPermission_allowSudo - The user is allowed to use sudo to elevate privileges.

$sel:level:SetPermission', setPermission_level - The user's permission level, which must be set to one of the following strings. You cannot set your own permissions level.

deny

show

deploy

manage

iam_only

For more information about the permissions associated with these levels, see Managing User Permissions.

$sel:allowSsh:SetPermission', setPermission_allowSsh - The user is allowed to use SSH to communicate with the instance.

$sel:stackId:SetPermission', setPermission_stackId - The stack ID.

$sel:iamUserArn:SetPermission', setPermission_iamUserArn - The user's IAM ARN. This can also be a federated user's ARN.

Request Lenses

setPermission_allowSudo :: Lens' SetPermission (Maybe Bool)

The user is allowed to use sudo to elevate privileges.

setPermission_level :: Lens' SetPermission (Maybe Text)

The user's permission level, which must be set to one of the following strings. You cannot set your own permissions level.

deny

show

deploy

manage

iam_only

For more information about the permissions associated with these levels, see Managing User Permissions.

setPermission_allowSsh :: Lens' SetPermission (Maybe Bool)

The user is allowed to use SSH to communicate with the instance.

setPermission_stackId :: Lens' SetPermission Text

The stack ID.

setPermission_iamUserArn :: Lens' SetPermission Text

The user's IAM ARN. This can also be a federated user's ARN.

Destructuring the Response

data SetPermissionResponse

See: newSetPermissionResponse smart constructor.

Constructors

SetPermissionResponse'

newSetPermissionResponse :: SetPermissionResponse

Create a value of SetPermissionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.