Network.AWS.OpsWorks.Types.AutoScalingThresholds

Description

Documentation

data AutoScalingThresholds

Describes a load-based auto scaling upscaling or downscaling threshold configuration, which specifies when AWS OpsWorks Stacks starts or stops load-based instances.

See: newAutoScalingThresholds smart constructor.

Constructors

AutoScalingThresholds'

Fields

instanceCount :: Maybe Int

The number of instances to add or remove when the load exceeds a threshold.

ignoreMetricsTime :: Maybe Natural

The amount of time (in minutes) after a scaling event occurs that AWS OpsWorks Stacks should ignore metrics and suppress additional scaling events. For example, AWS OpsWorks Stacks adds new instances following an upscaling event but the instances won't start reducing the load until they have been booted and configured. There is no point in raising additional scaling events during that operation, which typically takes several minutes. IgnoreMetricsTime allows you to direct AWS OpsWorks Stacks to suppress scaling events long enough to get the new instances online.

loadThreshold :: Maybe Double

The load threshold. A value of -1 disables the threshold. For more information about how load is computed, see Load (computing).

thresholdsWaitTime :: Maybe Natural

The amount of time, in minutes, that the load must exceed a threshold before more instances are added or removed.

alarms :: Maybe [Text]

Custom Cloudwatch auto scaling alarms, to be used as thresholds. This parameter takes a list of up to five alarm names, which are case sensitive and must be in the same region as the stack.

To use custom alarms, you must update your service role to allow cloudwatch:DescribeAlarms. You can either have AWS OpsWorks Stacks update the role for you when you first use this feature or you can edit the role manually. For more information, see Allowing AWS OpsWorks Stacks to Act on Your Behalf.

memoryThreshold :: Maybe Double

The memory utilization threshold, as a percent of the available memory. A value of -1 disables the threshold.

cpuThreshold :: Maybe Double

The CPU utilization threshold, as a percent of the available CPU. A value of -1 disables the threshold.

newAutoScalingThresholds :: AutoScalingThresholds

Create a value of AutoScalingThresholds with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceCount:AutoScalingThresholds', autoScalingThresholds_instanceCount - The number of instances to add or remove when the load exceeds a threshold.

$sel:ignoreMetricsTime:AutoScalingThresholds', autoScalingThresholds_ignoreMetricsTime - The amount of time (in minutes) after a scaling event occurs that AWS OpsWorks Stacks should ignore metrics and suppress additional scaling events. For example, AWS OpsWorks Stacks adds new instances following an upscaling event but the instances won't start reducing the load until they have been booted and configured. There is no point in raising additional scaling events during that operation, which typically takes several minutes. IgnoreMetricsTime allows you to direct AWS OpsWorks Stacks to suppress scaling events long enough to get the new instances online.

$sel:loadThreshold:AutoScalingThresholds', autoScalingThresholds_loadThreshold - The load threshold. A value of -1 disables the threshold. For more information about how load is computed, see Load (computing).

$sel:thresholdsWaitTime:AutoScalingThresholds', autoScalingThresholds_thresholdsWaitTime - The amount of time, in minutes, that the load must exceed a threshold before more instances are added or removed.

$sel:alarms:AutoScalingThresholds', autoScalingThresholds_alarms - Custom Cloudwatch auto scaling alarms, to be used as thresholds. This parameter takes a list of up to five alarm names, which are case sensitive and must be in the same region as the stack.

To use custom alarms, you must update your service role to allow cloudwatch:DescribeAlarms. You can either have AWS OpsWorks Stacks update the role for you when you first use this feature or you can edit the role manually. For more information, see Allowing AWS OpsWorks Stacks to Act on Your Behalf.

$sel:memoryThreshold:AutoScalingThresholds', autoScalingThresholds_memoryThreshold - The memory utilization threshold, as a percent of the available memory. A value of -1 disables the threshold.

$sel:cpuThreshold:AutoScalingThresholds', autoScalingThresholds_cpuThreshold - The CPU utilization threshold, as a percent of the available CPU. A value of -1 disables the threshold.

autoScalingThresholds_instanceCount :: Lens' AutoScalingThresholds (Maybe Int)

The number of instances to add or remove when the load exceeds a threshold.

autoScalingThresholds_ignoreMetricsTime :: Lens' AutoScalingThresholds (Maybe Natural)

The amount of time (in minutes) after a scaling event occurs that AWS OpsWorks Stacks should ignore metrics and suppress additional scaling events. For example, AWS OpsWorks Stacks adds new instances following an upscaling event but the instances won't start reducing the load until they have been booted and configured. There is no point in raising additional scaling events during that operation, which typically takes several minutes. IgnoreMetricsTime allows you to direct AWS OpsWorks Stacks to suppress scaling events long enough to get the new instances online.

autoScalingThresholds_loadThreshold :: Lens' AutoScalingThresholds (Maybe Double)

The load threshold. A value of -1 disables the threshold. For more information about how load is computed, see Load (computing).

autoScalingThresholds_thresholdsWaitTime :: Lens' AutoScalingThresholds (Maybe Natural)

The amount of time, in minutes, that the load must exceed a threshold before more instances are added or removed.

autoScalingThresholds_alarms :: Lens' AutoScalingThresholds (Maybe [Text])

Custom Cloudwatch auto scaling alarms, to be used as thresholds. This parameter takes a list of up to five alarm names, which are case sensitive and must be in the same region as the stack.

To use custom alarms, you must update your service role to allow cloudwatch:DescribeAlarms. You can either have AWS OpsWorks Stacks update the role for you when you first use this feature or you can edit the role manually. For more information, see Allowing AWS OpsWorks Stacks to Act on Your Behalf.

autoScalingThresholds_memoryThreshold :: Lens' AutoScalingThresholds (Maybe Double)

The memory utilization threshold, as a percent of the available memory. A value of -1 disables the threshold.

autoScalingThresholds_cpuThreshold :: Lens' AutoScalingThresholds (Maybe Double)

The CPU utilization threshold, as a percent of the available CPU. A value of -1 disables the threshold.