Network.AWS.OpsWorks.Types.Deployment

Description

Documentation

data Deployment

Describes a deployment of a stack or app.

See: newDeployment smart constructor.

Constructors

Deployment'

Fields

deploymentId :: Maybe Text

The deployment ID.

status :: Maybe Text

The deployment status:

running

successful

failed

command :: Maybe DeploymentCommand

Used to specify a stack or deployment command.

createdAt :: Maybe Text

Date when the deployment was created.

customJson :: Maybe Text

A string that contains user-defined custom JSON. It can be used to override the corresponding default stack configuration attribute values for stack or to pass data to recipes. The string should be in the following format:

"{\"key1\": \"value1\", \"key2\": \"value2\",...}"

For more information on custom JSON, see Use Custom JSON to Modify the Stack Configuration Attributes.

iamUserArn :: Maybe Text

The user's IAM ARN.

appId :: Maybe Text

The app ID.

instanceIds :: Maybe [Text]

The IDs of the target instances.

completedAt :: Maybe Text

Date when the deployment completed.

stackId :: Maybe Text

The stack ID.

comment :: Maybe Text

A user-defined comment.

duration :: Maybe Int

The deployment duration.

newDeployment :: Deployment

Create a value of Deployment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentId:Deployment', deployment_deploymentId - The deployment ID.

$sel:status:Deployment', deployment_status - The deployment status:

running

successful

failed

$sel:command:Deployment', deployment_command - Used to specify a stack or deployment command.

$sel:createdAt:Deployment', deployment_createdAt - Date when the deployment was created.

$sel:customJson:Deployment', deployment_customJson - A string that contains user-defined custom JSON. It can be used to override the corresponding default stack configuration attribute values for stack or to pass data to recipes. The string should be in the following format:

"{\"key1\": \"value1\", \"key2\": \"value2\",...}"

For more information on custom JSON, see Use Custom JSON to Modify the Stack Configuration Attributes.

$sel:iamUserArn:Deployment', deployment_iamUserArn - The user's IAM ARN.

$sel:appId:Deployment', deployment_appId - The app ID.

$sel:instanceIds:Deployment', deployment_instanceIds - The IDs of the target instances.

$sel:completedAt:Deployment', deployment_completedAt - Date when the deployment completed.

$sel:stackId:Deployment', deployment_stackId - The stack ID.

$sel:comment:Deployment', deployment_comment - A user-defined comment.

$sel:duration:Deployment', deployment_duration - The deployment duration.

deployment_deploymentId :: Lens' Deployment (Maybe Text)

The deployment ID.

deployment_status :: Lens' Deployment (Maybe Text)

The deployment status:

running

successful

failed

deployment_command :: Lens' Deployment (Maybe DeploymentCommand)

Used to specify a stack or deployment command.

deployment_createdAt :: Lens' Deployment (Maybe Text)

Date when the deployment was created.

deployment_customJson :: Lens' Deployment (Maybe Text)

A string that contains user-defined custom JSON. It can be used to override the corresponding default stack configuration attribute values for stack or to pass data to recipes. The string should be in the following format:

"{\"key1\": \"value1\", \"key2\": \"value2\",...}"

For more information on custom JSON, see Use Custom JSON to Modify the Stack Configuration Attributes.

deployment_iamUserArn :: Lens' Deployment (Maybe Text)

The user's IAM ARN.

deployment_appId :: Lens' Deployment (Maybe Text)

The app ID.

deployment_instanceIds :: Lens' Deployment (Maybe [Text])

The IDs of the target instances.

deployment_completedAt :: Lens' Deployment (Maybe Text)

Date when the deployment completed.

deployment_stackId :: Lens' Deployment (Maybe Text)

The stack ID.

deployment_comment :: Lens' Deployment (Maybe Text)

A user-defined comment.

deployment_duration :: Lens' Deployment (Maybe Int)

The deployment duration.