Network.AWS.OpsWorks.Types.ElasticLoadBalancer

Description

Documentation

data ElasticLoadBalancer

Describes an Elastic Load Balancing instance.

See: newElasticLoadBalancer smart constructor.

Constructors

ElasticLoadBalancer'

Fields

subnetIds :: Maybe [Text]

A list of subnet IDs, if the stack is running in a VPC.

vpcId :: Maybe Text

The VPC ID.

availabilityZones :: Maybe [Text]

A list of Availability Zones.

region :: Maybe Text

The instance's AWS region.

elasticLoadBalancerName :: Maybe Text

The Elastic Load Balancing instance's name.

stackId :: Maybe Text

The ID of the stack that the instance is associated with.

ec2InstanceIds :: Maybe [Text]

A list of the EC2 instances that the Elastic Load Balancing instance is managing traffic for.

layerId :: Maybe Text

The ID of the layer that the instance is attached to.

dnsName :: Maybe Text

The instance's public DNS name.

newElasticLoadBalancer :: ElasticLoadBalancer

Create a value of ElasticLoadBalancer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subnetIds:ElasticLoadBalancer', elasticLoadBalancer_subnetIds - A list of subnet IDs, if the stack is running in a VPC.

$sel:vpcId:ElasticLoadBalancer', elasticLoadBalancer_vpcId - The VPC ID.

$sel:availabilityZones:ElasticLoadBalancer', elasticLoadBalancer_availabilityZones - A list of Availability Zones.

$sel:region:ElasticLoadBalancer', elasticLoadBalancer_region - The instance's AWS region.

$sel:elasticLoadBalancerName:ElasticLoadBalancer', elasticLoadBalancer_elasticLoadBalancerName - The Elastic Load Balancing instance's name.

$sel:stackId:ElasticLoadBalancer', elasticLoadBalancer_stackId - The ID of the stack that the instance is associated with.

$sel:ec2InstanceIds:ElasticLoadBalancer', elasticLoadBalancer_ec2InstanceIds - A list of the EC2 instances that the Elastic Load Balancing instance is managing traffic for.

$sel:layerId:ElasticLoadBalancer', elasticLoadBalancer_layerId - The ID of the layer that the instance is attached to.

$sel:dnsName:ElasticLoadBalancer', elasticLoadBalancer_dnsName - The instance's public DNS name.

elasticLoadBalancer_subnetIds :: Lens' ElasticLoadBalancer (Maybe [Text])

A list of subnet IDs, if the stack is running in a VPC.

elasticLoadBalancer_vpcId :: Lens' ElasticLoadBalancer (Maybe Text)

The VPC ID.

elasticLoadBalancer_availabilityZones :: Lens' ElasticLoadBalancer (Maybe [Text])

A list of Availability Zones.

elasticLoadBalancer_region :: Lens' ElasticLoadBalancer (Maybe Text)

The instance's AWS region.

elasticLoadBalancer_elasticLoadBalancerName :: Lens' ElasticLoadBalancer (Maybe Text)

The Elastic Load Balancing instance's name.

elasticLoadBalancer_stackId :: Lens' ElasticLoadBalancer (Maybe Text)

The ID of the stack that the instance is associated with.

elasticLoadBalancer_ec2InstanceIds :: Lens' ElasticLoadBalancer (Maybe [Text])

A list of the EC2 instances that the Elastic Load Balancing instance is managing traffic for.

elasticLoadBalancer_layerId :: Lens' ElasticLoadBalancer (Maybe Text)

The ID of the layer that the instance is attached to.

elasticLoadBalancer_dnsName :: Lens' ElasticLoadBalancer (Maybe Text)

The instance's public DNS name.