Network.AWS.OpsWorks.Types.Instance

Description

Documentation

data Instance

Describes an instance.

See: newInstance smart constructor.

Constructors

Instance'

Fields

privateDns :: Maybe Text

The instance's private DNS name.

reportedAgentVersion :: Maybe Text

The instance's reported AWS OpsWorks Stacks agent version.

instanceId :: Maybe Text

The instance ID.

status :: Maybe Text

The instance status:

booting

connection_lost

online

pending

rebooting

requested

running_setup

setup_failed

shutting_down

start_failed

stop_failed

stopped

stopping

terminated

terminating

privateIp :: Maybe Text

The instance's private IP address.

installUpdatesOnBoot :: Maybe Bool

Whether to install operating system and package updates when the instance boots. The default value is true. If this value is set to false, you must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or by manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

virtualizationType :: Maybe VirtualizationType

The instance's virtualization type: paravirtual or hvm.

instanceProfileArn :: Maybe Text

The ARN of the instance's IAM profile. For more information about IAM ARNs, see Using Identifiers.

platform :: Maybe Text

The instance's platform.

hostname :: Maybe Text

The instance host name.

sshHostRsaKeyFingerprint :: Maybe Text

The SSH key's RSA fingerprint.

securityGroupIds :: Maybe [Text]

An array containing the instance security group IDs.

ecsClusterArn :: Maybe Text

For container instances, the Amazon ECS cluster's ARN.

arn :: Maybe Text

The instance's Amazon Resource Number (ARN).

createdAt :: Maybe Text

The time that the instance was created.

ec2InstanceId :: Maybe Text

The ID of the associated Amazon EC2 instance.

sshKeyName :: Maybe Text

The instance's Amazon EC2 key-pair name.

agentVersion :: Maybe Text

The agent version. This parameter is set to INHERIT if the instance inherits the default stack setting or to a a version number for a fixed agent version.

rootDeviceVolumeId :: Maybe Text

The root device volume ID.

subnetId :: Maybe Text

The instance's subnet ID; applicable only if the stack is running in a VPC.

infrastructureClass :: Maybe Text

For registered instances, the infrastructure class: ec2 or on-premises.

sshHostDsaKeyFingerprint :: Maybe Text

The SSH key's Deep Security Agent (DSA) fingerprint.

instanceType :: Maybe Text

The instance type, such as t2.micro.

ebsOptimized :: Maybe Bool

Whether this is an Amazon EBS-optimized instance.

elasticIp :: Maybe Text

The instance Elastic IP address .

os :: Maybe Text

The instance's operating system.

availabilityZone :: Maybe Text

The instance Availability Zone. For more information, see Regions and Endpoints.

lastServiceErrorId :: Maybe Text

The ID of the last service error. For more information, call DescribeServiceErrors.

tenancy :: Maybe Text

The instance's tenancy option, such as dedicated or host.

autoScalingType :: Maybe AutoScalingType

For load-based or time-based instances, the type.

layerIds :: Maybe [Text]

An array containing the instance layer IDs.

architecture :: Maybe Architecture

The instance architecture: "i386" or "x86_64".

publicDns :: Maybe Text

The instance public DNS name.

amiId :: Maybe Text

A custom AMI ID to be used to create the instance. For more information, see Instances

publicIp :: Maybe Text

The instance public IP address.

reportedOs :: Maybe ReportedOs

For registered instances, the reported operating system.

registeredBy :: Maybe Text

For registered instances, who performed the registration.

stackId :: Maybe Text

The stack ID.

rootDeviceType :: Maybe RootDeviceType

The instance's root device type. For more information, see Storage for the Root Device.

ecsContainerInstanceArn :: Maybe Text

For container instances, the instance's ARN.

blockDeviceMappings :: Maybe [BlockDeviceMapping]

An array of BlockDeviceMapping objects that specify the instance's block device mappings.

newInstance :: Instance

Create a value of Instance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:privateDns:Instance', instance_privateDns - The instance's private DNS name.

$sel:reportedAgentVersion:Instance', instance_reportedAgentVersion - The instance's reported AWS OpsWorks Stacks agent version.

$sel:instanceId:Instance', instance_instanceId - The instance ID.

$sel:status:Instance', instance_status - The instance status:

booting

connection_lost

online

pending

rebooting

requested

running_setup

setup_failed

shutting_down

start_failed

stop_failed

stopped

stopping

terminated

terminating

$sel:privateIp:Instance', instance_privateIp - The instance's private IP address.

$sel:installUpdatesOnBoot:Instance', instance_installUpdatesOnBoot - Whether to install operating system and package updates when the instance boots. The default value is true. If this value is set to false, you must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or by manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

$sel:virtualizationType:Instance', instance_virtualizationType - The instance's virtualization type: paravirtual or hvm.

$sel:instanceProfileArn:Instance', instance_instanceProfileArn - The ARN of the instance's IAM profile. For more information about IAM ARNs, see Using Identifiers.

$sel:platform:Instance', instance_platform - The instance's platform.

$sel:hostname:Instance', instance_hostname - The instance host name.

$sel:sshHostRsaKeyFingerprint:Instance', instance_sshHostRsaKeyFingerprint - The SSH key's RSA fingerprint.

$sel:securityGroupIds:Instance', instance_securityGroupIds - An array containing the instance security group IDs.

$sel:ecsClusterArn:Instance', instance_ecsClusterArn - For container instances, the Amazon ECS cluster's ARN.

$sel:arn:Instance', instance_arn - The instance's Amazon Resource Number (ARN).

$sel:createdAt:Instance', instance_createdAt - The time that the instance was created.

$sel:ec2InstanceId:Instance', instance_ec2InstanceId - The ID of the associated Amazon EC2 instance.

$sel:sshKeyName:Instance', instance_sshKeyName - The instance's Amazon EC2 key-pair name.

$sel:agentVersion:Instance', instance_agentVersion - The agent version. This parameter is set to INHERIT if the instance inherits the default stack setting or to a a version number for a fixed agent version.

$sel:rootDeviceVolumeId:Instance', instance_rootDeviceVolumeId - The root device volume ID.

$sel:subnetId:Instance', instance_subnetId - The instance's subnet ID; applicable only if the stack is running in a VPC.

$sel:infrastructureClass:Instance', instance_infrastructureClass - For registered instances, the infrastructure class: ec2 or on-premises.

$sel:sshHostDsaKeyFingerprint:Instance', instance_sshHostDsaKeyFingerprint - The SSH key's Deep Security Agent (DSA) fingerprint.

$sel:instanceType:Instance', instance_instanceType - The instance type, such as t2.micro.

$sel:ebsOptimized:Instance', instance_ebsOptimized - Whether this is an Amazon EBS-optimized instance.

$sel:elasticIp:Instance', instance_elasticIp - The instance Elastic IP address .

$sel:os:Instance', instance_os - The instance's operating system.

$sel:availabilityZone:Instance', instance_availabilityZone - The instance Availability Zone. For more information, see Regions and Endpoints.

$sel:lastServiceErrorId:Instance', instance_lastServiceErrorId - The ID of the last service error. For more information, call DescribeServiceErrors.

$sel:tenancy:Instance', instance_tenancy - The instance's tenancy option, such as dedicated or host.

$sel:autoScalingType:Instance', instance_autoScalingType - For load-based or time-based instances, the type.

$sel:layerIds:Instance', instance_layerIds - An array containing the instance layer IDs.

$sel:architecture:Instance', instance_architecture - The instance architecture: "i386" or "x86_64".

$sel:publicDns:Instance', instance_publicDns - The instance public DNS name.

$sel:amiId:Instance', instance_amiId - A custom AMI ID to be used to create the instance. For more information, see Instances

$sel:publicIp:Instance', instance_publicIp - The instance public IP address.

$sel:reportedOs:Instance', instance_reportedOs - For registered instances, the reported operating system.

$sel:registeredBy:Instance', instance_registeredBy - For registered instances, who performed the registration.

$sel:stackId:Instance', instance_stackId - The stack ID.

$sel:rootDeviceType:Instance', instance_rootDeviceType - The instance's root device type. For more information, see Storage for the Root Device.

$sel:ecsContainerInstanceArn:Instance', instance_ecsContainerInstanceArn - For container instances, the instance's ARN.

$sel:blockDeviceMappings:Instance', instance_blockDeviceMappings - An array of BlockDeviceMapping objects that specify the instance's block device mappings.

instance_privateDns :: Lens' Instance (Maybe Text)

The instance's private DNS name.

instance_reportedAgentVersion :: Lens' Instance (Maybe Text)

The instance's reported AWS OpsWorks Stacks agent version.

instance_instanceId :: Lens' Instance (Maybe Text)

The instance ID.

instance_status :: Lens' Instance (Maybe Text)

The instance status:

booting

connection_lost

online

pending

rebooting

requested

running_setup

setup_failed

shutting_down

start_failed

stop_failed

stopped

stopping

terminated

terminating

instance_privateIp :: Lens' Instance (Maybe Text)

The instance's private IP address.

instance_installUpdatesOnBoot :: Lens' Instance (Maybe Bool)

Whether to install operating system and package updates when the instance boots. The default value is true. If this value is set to false, you must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or by manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

instance_virtualizationType :: Lens' Instance (Maybe VirtualizationType)

The instance's virtualization type: paravirtual or hvm.

instance_instanceProfileArn :: Lens' Instance (Maybe Text)

The ARN of the instance's IAM profile. For more information about IAM ARNs, see Using Identifiers.

instance_platform :: Lens' Instance (Maybe Text)

The instance's platform.

instance_hostname :: Lens' Instance (Maybe Text)

The instance host name.

instance_sshHostRsaKeyFingerprint :: Lens' Instance (Maybe Text)

The SSH key's RSA fingerprint.

instance_securityGroupIds :: Lens' Instance (Maybe [Text])

An array containing the instance security group IDs.

instance_ecsClusterArn :: Lens' Instance (Maybe Text)

For container instances, the Amazon ECS cluster's ARN.

instance_arn :: Lens' Instance (Maybe Text)

The instance's Amazon Resource Number (ARN).

instance_createdAt :: Lens' Instance (Maybe Text)

The time that the instance was created.

instance_ec2InstanceId :: Lens' Instance (Maybe Text)

The ID of the associated Amazon EC2 instance.

instance_sshKeyName :: Lens' Instance (Maybe Text)

The instance's Amazon EC2 key-pair name.

instance_agentVersion :: Lens' Instance (Maybe Text)

The agent version. This parameter is set to INHERIT if the instance inherits the default stack setting or to a a version number for a fixed agent version.

instance_rootDeviceVolumeId :: Lens' Instance (Maybe Text)

The root device volume ID.

instance_subnetId :: Lens' Instance (Maybe Text)

The instance's subnet ID; applicable only if the stack is running in a VPC.

instance_infrastructureClass :: Lens' Instance (Maybe Text)

For registered instances, the infrastructure class: ec2 or on-premises.

instance_sshHostDsaKeyFingerprint :: Lens' Instance (Maybe Text)

The SSH key's Deep Security Agent (DSA) fingerprint.

instance_instanceType :: Lens' Instance (Maybe Text)

The instance type, such as t2.micro.

instance_ebsOptimized :: Lens' Instance (Maybe Bool)

Whether this is an Amazon EBS-optimized instance.

instance_elasticIp :: Lens' Instance (Maybe Text)

The instance Elastic IP address .

instance_os :: Lens' Instance (Maybe Text)

The instance's operating system.

instance_availabilityZone :: Lens' Instance (Maybe Text)

The instance Availability Zone. For more information, see Regions and Endpoints.

instance_lastServiceErrorId :: Lens' Instance (Maybe Text)

The ID of the last service error. For more information, call DescribeServiceErrors.

instance_tenancy :: Lens' Instance (Maybe Text)

The instance's tenancy option, such as dedicated or host.

instance_autoScalingType :: Lens' Instance (Maybe AutoScalingType)

For load-based or time-based instances, the type.

instance_layerIds :: Lens' Instance (Maybe [Text])

An array containing the instance layer IDs.

instance_architecture :: Lens' Instance (Maybe Architecture)

The instance architecture: "i386" or "x86_64".

instance_publicDns :: Lens' Instance (Maybe Text)

The instance public DNS name.

instance_amiId :: Lens' Instance (Maybe Text)

A custom AMI ID to be used to create the instance. For more information, see Instances

instance_publicIp :: Lens' Instance (Maybe Text)

The instance public IP address.

instance_reportedOs :: Lens' Instance (Maybe ReportedOs)

For registered instances, the reported operating system.

instance_registeredBy :: Lens' Instance (Maybe Text)

For registered instances, who performed the registration.

instance_stackId :: Lens' Instance (Maybe Text)

The stack ID.

instance_rootDeviceType :: Lens' Instance (Maybe RootDeviceType)

The instance's root device type. For more information, see Storage for the Root Device.

instance_ecsContainerInstanceArn :: Lens' Instance (Maybe Text)

For container instances, the instance's ARN.

instance_blockDeviceMappings :: Lens' Instance (Maybe [BlockDeviceMapping])

An array of BlockDeviceMapping objects that specify the instance's block device mappings.