Network.AWS.OpsWorks.Types.InstancesCount

Description

Documentation

data InstancesCount

Describes how many instances a stack has for each status.

See: newInstancesCount smart constructor.

Constructors

InstancesCount'

Fields

terminating :: Maybe Int

The number of instances with terminating status.

pending :: Maybe Int

The number of instances with pending status.

online :: Maybe Int

The number of instances with online status.

unassigning :: Maybe Int

The number of instances in the Unassigning state.

deregistering :: Maybe Int

The number of instances in the Deregistering state.

runningSetup :: Maybe Int

The number of instances with running_setup status.

requested :: Maybe Int

The number of instances with requested status.

stopFailed :: Maybe Int

The number of instances with stop_failed status.

booting :: Maybe Int

The number of instances with booting status.

stopped :: Maybe Int

The number of instances with stopped status.

rebooting :: Maybe Int

The number of instances with rebooting status.

assigning :: Maybe Int

The number of instances in the Assigning state.

shuttingDown :: Maybe Int

The number of instances with shutting_down status.

setupFailed :: Maybe Int

The number of instances with setup_failed status.

connectionLost :: Maybe Int

The number of instances with connection_lost status.

terminated :: Maybe Int

The number of instances with terminated status.

stopping :: Maybe Int

The number of instances with stopping status.

registered :: Maybe Int

The number of instances in the Registered state.

startFailed :: Maybe Int

The number of instances with start_failed status.

registering :: Maybe Int

The number of instances in the Registering state.

newInstancesCount :: InstancesCount

Create a value of InstancesCount with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:terminating:InstancesCount', instancesCount_terminating - The number of instances with terminating status.

$sel:pending:InstancesCount', instancesCount_pending - The number of instances with pending status.

$sel:online:InstancesCount', instancesCount_online - The number of instances with online status.

$sel:unassigning:InstancesCount', instancesCount_unassigning - The number of instances in the Unassigning state.

$sel:deregistering:InstancesCount', instancesCount_deregistering - The number of instances in the Deregistering state.

$sel:runningSetup:InstancesCount', instancesCount_runningSetup - The number of instances with running_setup status.

$sel:requested:InstancesCount', instancesCount_requested - The number of instances with requested status.

$sel:stopFailed:InstancesCount', instancesCount_stopFailed - The number of instances with stop_failed status.

$sel:booting:InstancesCount', instancesCount_booting - The number of instances with booting status.

$sel:stopped:InstancesCount', instancesCount_stopped - The number of instances with stopped status.

$sel:rebooting:InstancesCount', instancesCount_rebooting - The number of instances with rebooting status.

$sel:assigning:InstancesCount', instancesCount_assigning - The number of instances in the Assigning state.

$sel:shuttingDown:InstancesCount', instancesCount_shuttingDown - The number of instances with shutting_down status.

$sel:setupFailed:InstancesCount', instancesCount_setupFailed - The number of instances with setup_failed status.

$sel:connectionLost:InstancesCount', instancesCount_connectionLost - The number of instances with connection_lost status.

$sel:terminated:InstancesCount', instancesCount_terminated - The number of instances with terminated status.

$sel:stopping:InstancesCount', instancesCount_stopping - The number of instances with stopping status.

$sel:registered:InstancesCount', instancesCount_registered - The number of instances in the Registered state.

$sel:startFailed:InstancesCount', instancesCount_startFailed - The number of instances with start_failed status.

$sel:registering:InstancesCount', instancesCount_registering - The number of instances in the Registering state.

instancesCount_terminating :: Lens' InstancesCount (Maybe Int)

The number of instances with terminating status.

instancesCount_pending :: Lens' InstancesCount (Maybe Int)

The number of instances with pending status.

instancesCount_online :: Lens' InstancesCount (Maybe Int)

The number of instances with online status.

instancesCount_unassigning :: Lens' InstancesCount (Maybe Int)

The number of instances in the Unassigning state.

instancesCount_deregistering :: Lens' InstancesCount (Maybe Int)

The number of instances in the Deregistering state.

instancesCount_runningSetup :: Lens' InstancesCount (Maybe Int)

The number of instances with running_setup status.

instancesCount_requested :: Lens' InstancesCount (Maybe Int)

The number of instances with requested status.

instancesCount_stopFailed :: Lens' InstancesCount (Maybe Int)

The number of instances with stop_failed status.

instancesCount_booting :: Lens' InstancesCount (Maybe Int)

The number of instances with booting status.

instancesCount_stopped :: Lens' InstancesCount (Maybe Int)

The number of instances with stopped status.

instancesCount_rebooting :: Lens' InstancesCount (Maybe Int)

The number of instances with rebooting status.

instancesCount_assigning :: Lens' InstancesCount (Maybe Int)

The number of instances in the Assigning state.

instancesCount_shuttingDown :: Lens' InstancesCount (Maybe Int)

The number of instances with shutting_down status.

instancesCount_setupFailed :: Lens' InstancesCount (Maybe Int)

The number of instances with setup_failed status.

instancesCount_connectionLost :: Lens' InstancesCount (Maybe Int)

The number of instances with connection_lost status.

instancesCount_terminated :: Lens' InstancesCount (Maybe Int)

The number of instances with terminated status.

instancesCount_stopping :: Lens' InstancesCount (Maybe Int)

The number of instances with stopping status.

instancesCount_registered :: Lens' InstancesCount (Maybe Int)

The number of instances in the Registered state.

instancesCount_startFailed :: Lens' InstancesCount (Maybe Int)

The number of instances with start_failed status.

instancesCount_registering :: Lens' InstancesCount (Maybe Int)

The number of instances in the Registering state.