Network.AWS.OpsWorks.Types.OperatingSystem

Description

Documentation

data OperatingSystem

Describes supported operating systems in AWS OpsWorks Stacks.

See: newOperatingSystem smart constructor.

Constructors

OperatingSystem'

Fields

reportedVersion :: Maybe Text

The version of the operating system, including the release and edition, if applicable.

supported :: Maybe Bool

Indicates that an operating system is not supported for new instances.

name :: Maybe Text

The name of the operating system, such as Amazon Linux 2018.03.

id :: Maybe Text

The ID of a supported operating system, such as Amazon Linux 2018.03.

configurationManagers :: Maybe [OperatingSystemConfigurationManager]

Supported configuration manager name and versions for an AWS OpsWorks Stacks operating system.

type' :: Maybe Text

The type of a supported operating system, either Linux or Windows.

reportedName :: Maybe Text

A short name for the operating system manufacturer.

newOperatingSystem :: OperatingSystem

Create a value of OperatingSystem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportedVersion:OperatingSystem', operatingSystem_reportedVersion - The version of the operating system, including the release and edition, if applicable.

$sel:supported:OperatingSystem', operatingSystem_supported - Indicates that an operating system is not supported for new instances.

$sel:name:OperatingSystem', operatingSystem_name - The name of the operating system, such as Amazon Linux 2018.03.

$sel:id:OperatingSystem', operatingSystem_id - The ID of a supported operating system, such as Amazon Linux 2018.03.

$sel:configurationManagers:OperatingSystem', operatingSystem_configurationManagers - Supported configuration manager name and versions for an AWS OpsWorks Stacks operating system.

$sel:type':OperatingSystem', operatingSystem_type - The type of a supported operating system, either Linux or Windows.

$sel:reportedName:OperatingSystem', operatingSystem_reportedName - A short name for the operating system manufacturer.

operatingSystem_reportedVersion :: Lens' OperatingSystem (Maybe Text)

The version of the operating system, including the release and edition, if applicable.

operatingSystem_supported :: Lens' OperatingSystem (Maybe Bool)

Indicates that an operating system is not supported for new instances.

operatingSystem_name :: Lens' OperatingSystem (Maybe Text)

The name of the operating system, such as Amazon Linux 2018.03.

operatingSystem_id :: Lens' OperatingSystem (Maybe Text)

The ID of a supported operating system, such as Amazon Linux 2018.03.

operatingSystem_configurationManagers :: Lens' OperatingSystem (Maybe [OperatingSystemConfigurationManager])

Supported configuration manager name and versions for an AWS OpsWorks Stacks operating system.

operatingSystem_type :: Lens' OperatingSystem (Maybe Text)

The type of a supported operating system, either Linux or Windows.

operatingSystem_reportedName :: Lens' OperatingSystem (Maybe Text)

A short name for the operating system manufacturer.