Network.AWS.OpsWorks.Types.RdsDbInstance

Description

Documentation

data RdsDbInstance

Describes an Amazon RDS instance.

See: newRdsDbInstance smart constructor.

Constructors

RdsDbInstance'

Fields

rdsDbInstanceArn :: Maybe Text

The instance's ARN.

dbUser :: Maybe Text

The master user name.

missingOnRds :: Maybe Bool

Set to true if AWS OpsWorks Stacks is unable to discover the Amazon RDS instance. AWS OpsWorks Stacks attempts to discover the instance only once. If this value is set to true, you must deregister the instance, and then register it again.

engine :: Maybe Text

The instance's database engine.

address :: Maybe Text

The instance's address.

dbInstanceIdentifier :: Maybe Text

The DB instance identifier.

region :: Maybe Text

The instance's AWS region.

stackId :: Maybe Text

The ID of the stack with which the instance is registered.

dbPassword :: Maybe Text

AWS OpsWorks Stacks returns *****FILTERED***** instead of the actual value.

newRdsDbInstance :: RdsDbInstance

Create a value of RdsDbInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rdsDbInstanceArn:RdsDbInstance', rdsDbInstance_rdsDbInstanceArn - The instance's ARN.

$sel:dbUser:RdsDbInstance', rdsDbInstance_dbUser - The master user name.

$sel:missingOnRds:RdsDbInstance', rdsDbInstance_missingOnRds - Set to true if AWS OpsWorks Stacks is unable to discover the Amazon RDS instance. AWS OpsWorks Stacks attempts to discover the instance only once. If this value is set to true, you must deregister the instance, and then register it again.

$sel:engine:RdsDbInstance', rdsDbInstance_engine - The instance's database engine.

$sel:address:RdsDbInstance', rdsDbInstance_address - The instance's address.

$sel:dbInstanceIdentifier:RdsDbInstance', rdsDbInstance_dbInstanceIdentifier - The DB instance identifier.

$sel:region:RdsDbInstance', rdsDbInstance_region - The instance's AWS region.

$sel:stackId:RdsDbInstance', rdsDbInstance_stackId - The ID of the stack with which the instance is registered.

$sel:dbPassword:RdsDbInstance', rdsDbInstance_dbPassword - AWS OpsWorks Stacks returns *****FILTERED***** instead of the actual value.

rdsDbInstance_rdsDbInstanceArn :: Lens' RdsDbInstance (Maybe Text)

The instance's ARN.

rdsDbInstance_dbUser :: Lens' RdsDbInstance (Maybe Text)

The master user name.

rdsDbInstance_missingOnRds :: Lens' RdsDbInstance (Maybe Bool)

Set to true if AWS OpsWorks Stacks is unable to discover the Amazon RDS instance. AWS OpsWorks Stacks attempts to discover the instance only once. If this value is set to true, you must deregister the instance, and then register it again.

rdsDbInstance_engine :: Lens' RdsDbInstance (Maybe Text)

The instance's database engine.

rdsDbInstance_address :: Lens' RdsDbInstance (Maybe Text)

The instance's address.

rdsDbInstance_dbInstanceIdentifier :: Lens' RdsDbInstance (Maybe Text)

The DB instance identifier.

rdsDbInstance_region :: Lens' RdsDbInstance (Maybe Text)

The instance's AWS region.

rdsDbInstance_stackId :: Lens' RdsDbInstance (Maybe Text)

The ID of the stack with which the instance is registered.

rdsDbInstance_dbPassword :: Lens' RdsDbInstance (Maybe Text)

AWS OpsWorks Stacks returns *****FILTERED***** instead of the actual value.