Network.AWS.OpsWorks.Types.Source

Description

Documentation

data Source

Contains the information required to retrieve an app or cookbook from a repository. For more information, see Creating Apps or Custom Recipes and Cookbooks.

See: newSource smart constructor.

Constructors

Source'

Fields

url :: Maybe Text

The source URL. The following is an example of an Amazon S3 source URL: https://s3.amazonaws.com/opsworks-demo-bucket/opsworks_cookbook_demo.tar.gz.

username :: Maybe Text

This parameter depends on the repository type.

For Amazon S3 bundles, set Username to the appropriate IAM access key ID.

For HTTP bundles, Git repositories, and Subversion repositories, set Username to the user name.

sshKey :: Maybe Text

In requests, the repository's SSH key.

In responses, AWS OpsWorks Stacks returns *****FILTERED***** instead of the actual value.

password :: Maybe Text

When included in a request, the parameter depends on the repository type.

For Amazon S3 bundles, set Password to the appropriate IAM secret access key.

For HTTP bundles and Subversion repositories, set Password to the password.

For more information on how to safely handle IAM credentials, see https://docs.aws.amazon.com/general/latest/gr/aws-access-keys-best-practices.html.

In responses, AWS OpsWorks Stacks returns *****FILTERED***** instead of the actual value.

type' :: Maybe SourceType

The repository type.

revision :: Maybe Text

The application's version. AWS OpsWorks Stacks enables you to easily deploy new versions of an application. One of the simplest approaches is to have branches or revisions in your repository that represent different versions that can potentially be deployed.

newSource :: Source

Create a value of Source with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:url:Source', source_url - The source URL. The following is an example of an Amazon S3 source URL: https://s3.amazonaws.com/opsworks-demo-bucket/opsworks_cookbook_demo.tar.gz.

$sel:username:Source', source_username - This parameter depends on the repository type.

For Amazon S3 bundles, set Username to the appropriate IAM access key ID.

For HTTP bundles, Git repositories, and Subversion repositories, set Username to the user name.

$sel:sshKey:Source', source_sshKey - In requests, the repository's SSH key.

In responses, AWS OpsWorks Stacks returns *****FILTERED***** instead of the actual value.

$sel:password:Source', source_password - When included in a request, the parameter depends on the repository type.

For Amazon S3 bundles, set Password to the appropriate IAM secret access key.

For HTTP bundles and Subversion repositories, set Password to the password.

For more information on how to safely handle IAM credentials, see https://docs.aws.amazon.com/general/latest/gr/aws-access-keys-best-practices.html.

In responses, AWS OpsWorks Stacks returns *****FILTERED***** instead of the actual value.

$sel:type':Source', source_type - The repository type.

$sel:revision:Source', source_revision - The application's version. AWS OpsWorks Stacks enables you to easily deploy new versions of an application. One of the simplest approaches is to have branches or revisions in your repository that represent different versions that can potentially be deployed.

source_url :: Lens' Source (Maybe Text)

The source URL. The following is an example of an Amazon S3 source URL: https://s3.amazonaws.com/opsworks-demo-bucket/opsworks_cookbook_demo.tar.gz.

source_username :: Lens' Source (Maybe Text)

This parameter depends on the repository type.

For Amazon S3 bundles, set Username to the appropriate IAM access key ID.

For HTTP bundles, Git repositories, and Subversion repositories, set Username to the user name.

source_sshKey :: Lens' Source (Maybe Text)

In requests, the repository's SSH key.

In responses, AWS OpsWorks Stacks returns *****FILTERED***** instead of the actual value.

source_password :: Lens' Source (Maybe Text)

When included in a request, the parameter depends on the repository type.

For Amazon S3 bundles, set Password to the appropriate IAM secret access key.

For HTTP bundles and Subversion repositories, set Password to the password.

For more information on how to safely handle IAM credentials, see https://docs.aws.amazon.com/general/latest/gr/aws-access-keys-best-practices.html.

In responses, AWS OpsWorks Stacks returns *****FILTERED***** instead of the actual value.

source_type :: Lens' Source (Maybe SourceType)

The repository type.

source_revision :: Lens' Source (Maybe Text)

The application's version. AWS OpsWorks Stacks enables you to easily deploy new versions of an application. One of the simplest approaches is to have branches or revisions in your repository that represent different versions that can potentially be deployed.