Network.AWS.OpsWorks.Types.Stack

Description

Documentation

data Stack

Describes a stack.

See: newStack smart constructor.

Constructors

Stack'

Fields

defaultInstanceProfileArn :: Maybe Text

The ARN of an IAM profile that is the default profile for all of the stack's EC2 instances. For more information about IAM ARNs, see Using Identifiers.

serviceRoleArn :: Maybe Text

The stack AWS Identity and Access Management (IAM) role.

defaultRootDeviceType :: Maybe RootDeviceType

The default root device type. This value is used by default for all instances in the stack, but you can override it when you create an instance. For more information, see Storage for the Root Device.

arn :: Maybe Text

The stack's ARN.

createdAt :: Maybe Text

The date when the stack was created.

vpcId :: Maybe Text

The VPC ID; applicable only if the stack is running in a VPC.

chefConfiguration :: Maybe ChefConfiguration

A ChefConfiguration object that specifies whether to enable Berkshelf and the Berkshelf version. For more information, see Create a New Stack.

agentVersion :: Maybe Text

The agent version. This parameter is set to LATEST for auto-update. or a version number for a fixed agent version.

defaultSshKeyName :: Maybe Text

A default Amazon EC2 key pair for the stack's instances. You can override this value when you create or update an instance.

customJson :: Maybe Text

A JSON object that contains user-defined attributes to be added to the stack configuration and deployment attributes. You can use custom JSON to override the corresponding default stack configuration attribute values or to pass data to recipes. The string should be in the following format:

"{\"key1\": \"value1\", \"key2\": \"value2\",...}"

For more information on custom JSON, see Use Custom JSON to Modify the Stack Configuration Attributes.

customCookbooksSource :: Maybe Source

Contains the information required to retrieve an app or cookbook from a repository. For more information, see Adding Apps or Cookbooks and Recipes.

defaultAvailabilityZone :: Maybe Text

The stack's default Availability Zone. For more information, see Regions and Endpoints.

attributes :: Maybe (HashMap StackAttributesKeys (Maybe Text))

The stack's attributes.

name :: Maybe Text

The stack name.

defaultOs :: Maybe Text

The stack's default operating system.

useOpsworksSecurityGroups :: Maybe Bool

Whether the stack automatically associates the AWS OpsWorks Stacks built-in security groups with the stack's layers.

useCustomCookbooks :: Maybe Bool

Whether the stack uses custom cookbooks.

defaultSubnetId :: Maybe Text

The default subnet ID; applicable only if the stack is running in a VPC.

region :: Maybe Text

The stack AWS region, such as "ap-northeast-2". For more information about AWS regions, see Regions and Endpoints.

configurationManager :: Maybe StackConfigurationManager

The configuration manager.

stackId :: Maybe Text

The stack ID.

hostnameTheme :: Maybe Text

The stack host name theme, with spaces replaced by underscores.

newStack :: Stack

Create a value of Stack with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defaultInstanceProfileArn:Stack', stack_defaultInstanceProfileArn - The ARN of an IAM profile that is the default profile for all of the stack's EC2 instances. For more information about IAM ARNs, see Using Identifiers.

$sel:serviceRoleArn:Stack', stack_serviceRoleArn - The stack AWS Identity and Access Management (IAM) role.

$sel:defaultRootDeviceType:Stack', stack_defaultRootDeviceType - The default root device type. This value is used by default for all instances in the stack, but you can override it when you create an instance. For more information, see Storage for the Root Device.

$sel:arn:Stack', stack_arn - The stack's ARN.

$sel:createdAt:Stack', stack_createdAt - The date when the stack was created.

$sel:vpcId:Stack', stack_vpcId - The VPC ID; applicable only if the stack is running in a VPC.

$sel:chefConfiguration:Stack', stack_chefConfiguration - A ChefConfiguration object that specifies whether to enable Berkshelf and the Berkshelf version. For more information, see Create a New Stack.

$sel:agentVersion:Stack', stack_agentVersion - The agent version. This parameter is set to LATEST for auto-update. or a version number for a fixed agent version.

$sel:defaultSshKeyName:Stack', stack_defaultSshKeyName - A default Amazon EC2 key pair for the stack's instances. You can override this value when you create or update an instance.

$sel:customJson:Stack', stack_customJson - A JSON object that contains user-defined attributes to be added to the stack configuration and deployment attributes. You can use custom JSON to override the corresponding default stack configuration attribute values or to pass data to recipes. The string should be in the following format:

"{\"key1\": \"value1\", \"key2\": \"value2\",...}"

For more information on custom JSON, see Use Custom JSON to Modify the Stack Configuration Attributes.

$sel:customCookbooksSource:Stack', stack_customCookbooksSource - Contains the information required to retrieve an app or cookbook from a repository. For more information, see Adding Apps or Cookbooks and Recipes.

$sel:defaultAvailabilityZone:Stack', stack_defaultAvailabilityZone - The stack's default Availability Zone. For more information, see Regions and Endpoints.

$sel:attributes:Stack', stack_attributes - The stack's attributes.

$sel:name:Stack', stack_name - The stack name.

$sel:defaultOs:Stack', stack_defaultOs - The stack's default operating system.

$sel:useOpsworksSecurityGroups:Stack', stack_useOpsworksSecurityGroups - Whether the stack automatically associates the AWS OpsWorks Stacks built-in security groups with the stack's layers.

$sel:useCustomCookbooks:Stack', stack_useCustomCookbooks - Whether the stack uses custom cookbooks.

$sel:defaultSubnetId:Stack', stack_defaultSubnetId - The default subnet ID; applicable only if the stack is running in a VPC.

$sel:region:Stack', stack_region - The stack AWS region, such as "ap-northeast-2". For more information about AWS regions, see Regions and Endpoints.

$sel:configurationManager:Stack', stack_configurationManager - The configuration manager.

$sel:stackId:Stack', stack_stackId - The stack ID.

$sel:hostnameTheme:Stack', stack_hostnameTheme - The stack host name theme, with spaces replaced by underscores.

stack_defaultInstanceProfileArn :: Lens' Stack (Maybe Text)

The ARN of an IAM profile that is the default profile for all of the stack's EC2 instances. For more information about IAM ARNs, see Using Identifiers.

stack_serviceRoleArn :: Lens' Stack (Maybe Text)

The stack AWS Identity and Access Management (IAM) role.

stack_defaultRootDeviceType :: Lens' Stack (Maybe RootDeviceType)

The default root device type. This value is used by default for all instances in the stack, but you can override it when you create an instance. For more information, see Storage for the Root Device.

stack_arn :: Lens' Stack (Maybe Text)

The stack's ARN.

stack_createdAt :: Lens' Stack (Maybe Text)

The date when the stack was created.

stack_vpcId :: Lens' Stack (Maybe Text)

The VPC ID; applicable only if the stack is running in a VPC.

stack_chefConfiguration :: Lens' Stack (Maybe ChefConfiguration)

A ChefConfiguration object that specifies whether to enable Berkshelf and the Berkshelf version. For more information, see Create a New Stack.

stack_agentVersion :: Lens' Stack (Maybe Text)

The agent version. This parameter is set to LATEST for auto-update. or a version number for a fixed agent version.

stack_defaultSshKeyName :: Lens' Stack (Maybe Text)

A default Amazon EC2 key pair for the stack's instances. You can override this value when you create or update an instance.

stack_customJson :: Lens' Stack (Maybe Text)

A JSON object that contains user-defined attributes to be added to the stack configuration and deployment attributes. You can use custom JSON to override the corresponding default stack configuration attribute values or to pass data to recipes. The string should be in the following format:

"{\"key1\": \"value1\", \"key2\": \"value2\",...}"

For more information on custom JSON, see Use Custom JSON to Modify the Stack Configuration Attributes.

stack_customCookbooksSource :: Lens' Stack (Maybe Source)

Contains the information required to retrieve an app or cookbook from a repository. For more information, see Adding Apps or Cookbooks and Recipes.

stack_defaultAvailabilityZone :: Lens' Stack (Maybe Text)

The stack's default Availability Zone. For more information, see Regions and Endpoints.

stack_attributes :: Lens' Stack (Maybe (HashMap StackAttributesKeys (Maybe Text)))

The stack's attributes.

stack_name :: Lens' Stack (Maybe Text)

The stack name.

stack_defaultOs :: Lens' Stack (Maybe Text)

The stack's default operating system.

stack_useOpsworksSecurityGroups :: Lens' Stack (Maybe Bool)

Whether the stack automatically associates the AWS OpsWorks Stacks built-in security groups with the stack's layers.

stack_useCustomCookbooks :: Lens' Stack (Maybe Bool)

Whether the stack uses custom cookbooks.

stack_defaultSubnetId :: Lens' Stack (Maybe Text)

The default subnet ID; applicable only if the stack is running in a VPC.

stack_region :: Lens' Stack (Maybe Text)

The stack AWS region, such as "ap-northeast-2". For more information about AWS regions, see Regions and Endpoints.

stack_configurationManager :: Lens' Stack (Maybe StackConfigurationManager)

The configuration manager.

stack_stackId :: Lens' Stack (Maybe Text)

The stack ID.

stack_hostnameTheme :: Lens' Stack (Maybe Text)

The stack host name theme, with spaces replaced by underscores.