Network.AWS.OpsWorks.Types.Volume

Description

Documentation

data Volume

Describes an instance's Amazon EBS volume.

See: newVolume smart constructor.

Constructors

Volume'

Fields

instanceId :: Maybe Text

The instance ID.

status :: Maybe Text

The value returned by DescribeVolumes.

size :: Maybe Int

The volume size.

iops :: Maybe Int

For PIOPS volumes, the IOPS per disk.

device :: Maybe Text

The device name.

encrypted :: Maybe Bool

Specifies whether an Amazon EBS volume is encrypted. For more information, see Amazon EBS Encryption.

availabilityZone :: Maybe Text

The volume Availability Zone. For more information, see Regions and Endpoints.

name :: Maybe Text

The volume name.

raidArrayId :: Maybe Text

The RAID array ID.

volumeId :: Maybe Text

The volume ID.

region :: Maybe Text

The AWS region. For more information about AWS regions, see Regions and Endpoints.

volumeType :: Maybe Text

The volume type. For more information, see Amazon EBS Volume Types.

standard - Magnetic. Magnetic volumes must have a minimum size of 1 GiB and a maximum size of 1024 GiB.

io1 - Provisioned IOPS (SSD). PIOPS volumes must have a minimum size of 4 GiB and a maximum size of 16384 GiB.

gp2 - General Purpose (SSD). General purpose volumes must have a minimum size of 1 GiB and a maximum size of 16384 GiB.

st1 - Throughput Optimized hard disk drive (HDD). Throughput optimized HDD volumes must have a minimum size of 500 GiB and a maximum size of 16384 GiB.

sc1 - Cold HDD. Cold HDD volumes must have a minimum size of 500 GiB and a maximum size of 16384 GiB.

ec2VolumeId :: Maybe Text

The Amazon EC2 volume ID.

mountPoint :: Maybe Text

The volume mount point. For example, "/mnt/disk1".

newVolume :: Volume

Create a value of Volume with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceId:Volume', volume_instanceId - The instance ID.

$sel:status:Volume', volume_status - The value returned by DescribeVolumes.

$sel:size:Volume', volume_size - The volume size.

$sel:iops:Volume', volume_iops - For PIOPS volumes, the IOPS per disk.

$sel:device:Volume', volume_device - The device name.

$sel:encrypted:Volume', volume_encrypted - Specifies whether an Amazon EBS volume is encrypted. For more information, see Amazon EBS Encryption.

$sel:availabilityZone:Volume', volume_availabilityZone - The volume Availability Zone. For more information, see Regions and Endpoints.

$sel:name:Volume', volume_name - The volume name.

$sel:raidArrayId:Volume', volume_raidArrayId - The RAID array ID.

$sel:volumeId:Volume', volume_volumeId - The volume ID.

$sel:region:Volume', volume_region - The AWS region. For more information about AWS regions, see Regions and Endpoints.

$sel:volumeType:Volume', volume_volumeType - The volume type. For more information, see Amazon EBS Volume Types.

standard - Magnetic. Magnetic volumes must have a minimum size of 1 GiB and a maximum size of 1024 GiB.

io1 - Provisioned IOPS (SSD). PIOPS volumes must have a minimum size of 4 GiB and a maximum size of 16384 GiB.

gp2 - General Purpose (SSD). General purpose volumes must have a minimum size of 1 GiB and a maximum size of 16384 GiB.

st1 - Throughput Optimized hard disk drive (HDD). Throughput optimized HDD volumes must have a minimum size of 500 GiB and a maximum size of 16384 GiB.

sc1 - Cold HDD. Cold HDD volumes must have a minimum size of 500 GiB and a maximum size of 16384 GiB.

$sel:ec2VolumeId:Volume', volume_ec2VolumeId - The Amazon EC2 volume ID.

$sel:mountPoint:Volume', volume_mountPoint - The volume mount point. For example, "/mnt/disk1".

volume_instanceId :: Lens' Volume (Maybe Text)

The instance ID.

volume_status :: Lens' Volume (Maybe Text)

The value returned by DescribeVolumes.

volume_size :: Lens' Volume (Maybe Int)

The volume size.

volume_iops :: Lens' Volume (Maybe Int)

For PIOPS volumes, the IOPS per disk.

volume_device :: Lens' Volume (Maybe Text)

The device name.

volume_encrypted :: Lens' Volume (Maybe Bool)

Specifies whether an Amazon EBS volume is encrypted. For more information, see Amazon EBS Encryption.

volume_availabilityZone :: Lens' Volume (Maybe Text)

The volume Availability Zone. For more information, see Regions and Endpoints.

volume_name :: Lens' Volume (Maybe Text)

The volume name.

volume_raidArrayId :: Lens' Volume (Maybe Text)

The RAID array ID.

volume_volumeId :: Lens' Volume (Maybe Text)

The volume ID.

volume_region :: Lens' Volume (Maybe Text)

The AWS region. For more information about AWS regions, see Regions and Endpoints.

volume_volumeType :: Lens' Volume (Maybe Text)

The volume type. For more information, see Amazon EBS Volume Types.

standard - Magnetic. Magnetic volumes must have a minimum size of 1 GiB and a maximum size of 1024 GiB.

io1 - Provisioned IOPS (SSD). PIOPS volumes must have a minimum size of 4 GiB and a maximum size of 16384 GiB.

gp2 - General Purpose (SSD). General purpose volumes must have a minimum size of 1 GiB and a maximum size of 16384 GiB.

st1 - Throughput Optimized hard disk drive (HDD). Throughput optimized HDD volumes must have a minimum size of 500 GiB and a maximum size of 16384 GiB.

sc1 - Cold HDD. Cold HDD volumes must have a minimum size of 500 GiB and a maximum size of 16384 GiB.

volume_ec2VolumeId :: Lens' Volume (Maybe Text)

The Amazon EC2 volume ID.

volume_mountPoint :: Lens' Volume (Maybe Text)

The volume mount point. For example, "/mnt/disk1".