Network.AWS.OpsWorks.Types.VolumeConfiguration

Description

Documentation

data VolumeConfiguration

Describes an Amazon EBS volume configuration.

See: newVolumeConfiguration smart constructor.

Constructors

VolumeConfiguration'

Fields

iops :: Maybe Int

For PIOPS volumes, the IOPS per disk.

raidLevel :: Maybe Int

The volume RAID level.

encrypted :: Maybe Bool

Specifies whether an Amazon EBS volume is encrypted. For more information, see Amazon EBS Encryption.

volumeType :: Maybe Text

The volume type. For more information, see Amazon EBS Volume Types.

standard - Magnetic. Magnetic volumes must have a minimum size of 1 GiB and a maximum size of 1024 GiB.

io1 - Provisioned IOPS (SSD). PIOPS volumes must have a minimum size of 4 GiB and a maximum size of 16384 GiB.

gp2 - General Purpose (SSD). General purpose volumes must have a minimum size of 1 GiB and a maximum size of 16384 GiB.

st1 - Throughput Optimized hard disk drive (HDD). Throughput optimized HDD volumes must have a minimum size of 500 GiB and a maximum size of 16384 GiB.

sc1 - Cold HDD. Cold HDD volumes must have a minimum size of 500 GiB and a maximum size of 16384 GiB.

mountPoint :: Text

The volume mount point. For example "/dev/sdh".

numberOfDisks :: Int

The number of disks in the volume.

size :: Int

The volume size.

newVolumeConfiguration

Arguments

:: Text
$sel:mountPoint:VolumeConfiguration'

-> Int
$sel:numberOfDisks:VolumeConfiguration'

-> Int
$sel:size:VolumeConfiguration'

-> VolumeConfiguration

Create a value of VolumeConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:iops:VolumeConfiguration', volumeConfiguration_iops - For PIOPS volumes, the IOPS per disk.

$sel:raidLevel:VolumeConfiguration', volumeConfiguration_raidLevel - The volume RAID level.

$sel:encrypted:VolumeConfiguration', volumeConfiguration_encrypted - Specifies whether an Amazon EBS volume is encrypted. For more information, see Amazon EBS Encryption.

$sel:volumeType:VolumeConfiguration', volumeConfiguration_volumeType - The volume type. For more information, see Amazon EBS Volume Types.

standard - Magnetic. Magnetic volumes must have a minimum size of 1 GiB and a maximum size of 1024 GiB.

io1 - Provisioned IOPS (SSD). PIOPS volumes must have a minimum size of 4 GiB and a maximum size of 16384 GiB.

gp2 - General Purpose (SSD). General purpose volumes must have a minimum size of 1 GiB and a maximum size of 16384 GiB.

st1 - Throughput Optimized hard disk drive (HDD). Throughput optimized HDD volumes must have a minimum size of 500 GiB and a maximum size of 16384 GiB.

sc1 - Cold HDD. Cold HDD volumes must have a minimum size of 500 GiB and a maximum size of 16384 GiB.

$sel:mountPoint:VolumeConfiguration', volumeConfiguration_mountPoint - The volume mount point. For example "/dev/sdh".

$sel:numberOfDisks:VolumeConfiguration', volumeConfiguration_numberOfDisks - The number of disks in the volume.

$sel:size:VolumeConfiguration', volumeConfiguration_size - The volume size.

volumeConfiguration_iops :: Lens' VolumeConfiguration (Maybe Int)

For PIOPS volumes, the IOPS per disk.

volumeConfiguration_raidLevel :: Lens' VolumeConfiguration (Maybe Int)

The volume RAID level.

volumeConfiguration_encrypted :: Lens' VolumeConfiguration (Maybe Bool)

Specifies whether an Amazon EBS volume is encrypted. For more information, see Amazon EBS Encryption.

volumeConfiguration_volumeType :: Lens' VolumeConfiguration (Maybe Text)

The volume type. For more information, see Amazon EBS Volume Types.

standard - Magnetic. Magnetic volumes must have a minimum size of 1 GiB and a maximum size of 1024 GiB.

io1 - Provisioned IOPS (SSD). PIOPS volumes must have a minimum size of 4 GiB and a maximum size of 16384 GiB.

gp2 - General Purpose (SSD). General purpose volumes must have a minimum size of 1 GiB and a maximum size of 16384 GiB.

st1 - Throughput Optimized hard disk drive (HDD). Throughput optimized HDD volumes must have a minimum size of 500 GiB and a maximum size of 16384 GiB.

sc1 - Cold HDD. Cold HDD volumes must have a minimum size of 500 GiB and a maximum size of 16384 GiB.

volumeConfiguration_mountPoint :: Lens' VolumeConfiguration Text

The volume mount point. For example "/dev/sdh".

volumeConfiguration_numberOfDisks :: Lens' VolumeConfiguration Int

The number of disks in the volume.

volumeConfiguration_size :: Lens' VolumeConfiguration Int

The volume size.