Network.AWS.OpsWorks.Types.WeeklyAutoScalingSchedule

Description

Documentation

data WeeklyAutoScalingSchedule

Describes a time-based instance's auto scaling schedule. The schedule consists of a set of key-value pairs.

The key is the time period (a UTC hour) and must be an integer from 0 - 23.

The value indicates whether the instance should be online or offline for the specified period, and must be set to "on" or "off"

The default setting for all time periods is off, so you use the following parameters primarily to specify the online periods. You don't have to explicitly specify offline periods unless you want to change an online period to an offline period.

The following example specifies that the instance should be online for four hours, from UTC 1200 - 1600. It will be off for the remainder of the day.

 { "12":"on", "13":"on", "14":"on", "15":"on" }

See: newWeeklyAutoScalingSchedule smart constructor.

Constructors

WeeklyAutoScalingSchedule'

Fields

thursday :: Maybe (HashMap Text Text)

The schedule for Thursday.

wednesday :: Maybe (HashMap Text Text)

The schedule for Wednesday.

saturday :: Maybe (HashMap Text Text)

The schedule for Saturday.

monday :: Maybe (HashMap Text Text)

The schedule for Monday.

friday :: Maybe (HashMap Text Text)

The schedule for Friday.

sunday :: Maybe (HashMap Text Text)

The schedule for Sunday.

tuesday :: Maybe (HashMap Text Text)

The schedule for Tuesday.

newWeeklyAutoScalingSchedule :: WeeklyAutoScalingSchedule

Create a value of WeeklyAutoScalingSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:thursday:WeeklyAutoScalingSchedule', weeklyAutoScalingSchedule_thursday - The schedule for Thursday.

$sel:wednesday:WeeklyAutoScalingSchedule', weeklyAutoScalingSchedule_wednesday - The schedule for Wednesday.

$sel:saturday:WeeklyAutoScalingSchedule', weeklyAutoScalingSchedule_saturday - The schedule for Saturday.

$sel:monday:WeeklyAutoScalingSchedule', weeklyAutoScalingSchedule_monday - The schedule for Monday.

$sel:friday:WeeklyAutoScalingSchedule', weeklyAutoScalingSchedule_friday - The schedule for Friday.

$sel:sunday:WeeklyAutoScalingSchedule', weeklyAutoScalingSchedule_sunday - The schedule for Sunday.

$sel:tuesday:WeeklyAutoScalingSchedule', weeklyAutoScalingSchedule_tuesday - The schedule for Tuesday.

weeklyAutoScalingSchedule_thursday :: Lens' WeeklyAutoScalingSchedule (Maybe (HashMap Text Text))

The schedule for Thursday.

weeklyAutoScalingSchedule_wednesday :: Lens' WeeklyAutoScalingSchedule (Maybe (HashMap Text Text))

The schedule for Wednesday.

weeklyAutoScalingSchedule_saturday :: Lens' WeeklyAutoScalingSchedule (Maybe (HashMap Text Text))

The schedule for Saturday.

weeklyAutoScalingSchedule_monday :: Lens' WeeklyAutoScalingSchedule (Maybe (HashMap Text Text))

The schedule for Monday.

weeklyAutoScalingSchedule_friday :: Lens' WeeklyAutoScalingSchedule (Maybe (HashMap Text Text))

The schedule for Friday.

weeklyAutoScalingSchedule_sunday :: Lens' WeeklyAutoScalingSchedule (Maybe (HashMap Text Text))

The schedule for Sunday.

weeklyAutoScalingSchedule_tuesday :: Lens' WeeklyAutoScalingSchedule (Maybe (HashMap Text Text))

The schedule for Tuesday.