Network.AWS.OpsWorks.UpdateApp

Description

Updates a specified app.

Required Permissions: To use this action, an IAM user must have a Deploy or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Creating a Request

data UpdateApp

See: newUpdateApp smart constructor.

Constructors

UpdateApp'

Fields

sslConfiguration :: Maybe SslConfiguration

An SslConfiguration object with the SSL configuration.

environment :: Maybe [EnvironmentVariable]

An array of EnvironmentVariable objects that specify environment variables to be associated with the app. After you deploy the app, these variables are defined on the associated app server instances.For more information, see Environment Variables.

There is no specific limit on the number of environment variables. However, the size of the associated data structure - which includes the variables' names, values, and protected flag values - cannot exceed 20 KB. This limit should accommodate most if not all use cases. Exceeding it will cause an exception with the message, "Environment: is too large (maximum is 20 KB)."

If you have specified one or more environment variables, you cannot modify the stack's Chef version.

enableSsl :: Maybe Bool

Whether SSL is enabled for the app.

dataSources :: Maybe [DataSource]

The app's data sources.

appSource :: Maybe Source

A Source object that specifies the app repository.

attributes :: Maybe (HashMap AppAttributesKeys Text)

One or more user-defined key/value pairs to be added to the stack attributes.

name :: Maybe Text

The app name.

type' :: Maybe AppType

The app type.

domains :: Maybe [Text]

The app's virtual host settings, with multiple domains separated by commas. For example: 'www.example.com, example.com'

description :: Maybe Text

A description of the app.

appId :: Text

The app ID.

newUpdateApp

Arguments

:: Text
$sel:appId:UpdateApp'

-> UpdateApp

Create a value of UpdateApp with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sslConfiguration:UpdateApp', updateApp_sslConfiguration - An SslConfiguration object with the SSL configuration.

$sel:environment:UpdateApp', updateApp_environment - An array of EnvironmentVariable objects that specify environment variables to be associated with the app. After you deploy the app, these variables are defined on the associated app server instances.For more information, see Environment Variables.

There is no specific limit on the number of environment variables. However, the size of the associated data structure - which includes the variables' names, values, and protected flag values - cannot exceed 20 KB. This limit should accommodate most if not all use cases. Exceeding it will cause an exception with the message, "Environment: is too large (maximum is 20 KB)."

If you have specified one or more environment variables, you cannot modify the stack's Chef version.

$sel:enableSsl:UpdateApp', updateApp_enableSsl - Whether SSL is enabled for the app.

$sel:dataSources:UpdateApp', updateApp_dataSources - The app's data sources.

$sel:appSource:UpdateApp', updateApp_appSource - A Source object that specifies the app repository.

$sel:attributes:UpdateApp', updateApp_attributes - One or more user-defined key/value pairs to be added to the stack attributes.

$sel:name:UpdateApp', updateApp_name - The app name.

$sel:type':UpdateApp', updateApp_type - The app type.

$sel:domains:UpdateApp', updateApp_domains - The app's virtual host settings, with multiple domains separated by commas. For example: 'www.example.com, example.com'

$sel:description:UpdateApp', updateApp_description - A description of the app.

$sel:appId:UpdateApp', updateApp_appId - The app ID.

Request Lenses

updateApp_sslConfiguration :: Lens' UpdateApp (Maybe SslConfiguration)

An SslConfiguration object with the SSL configuration.

updateApp_environment :: Lens' UpdateApp (Maybe [EnvironmentVariable])

An array of EnvironmentVariable objects that specify environment variables to be associated with the app. After you deploy the app, these variables are defined on the associated app server instances.For more information, see Environment Variables.

There is no specific limit on the number of environment variables. However, the size of the associated data structure - which includes the variables' names, values, and protected flag values - cannot exceed 20 KB. This limit should accommodate most if not all use cases. Exceeding it will cause an exception with the message, "Environment: is too large (maximum is 20 KB)."

If you have specified one or more environment variables, you cannot modify the stack's Chef version.

updateApp_enableSsl :: Lens' UpdateApp (Maybe Bool)

Whether SSL is enabled for the app.

updateApp_dataSources :: Lens' UpdateApp (Maybe [DataSource])

The app's data sources.

updateApp_appSource :: Lens' UpdateApp (Maybe Source)

A Source object that specifies the app repository.

updateApp_attributes :: Lens' UpdateApp (Maybe (HashMap AppAttributesKeys Text))

One or more user-defined key/value pairs to be added to the stack attributes.

updateApp_name :: Lens' UpdateApp (Maybe Text)

The app name.

updateApp_type :: Lens' UpdateApp (Maybe AppType)

The app type.

updateApp_domains :: Lens' UpdateApp (Maybe [Text])

The app's virtual host settings, with multiple domains separated by commas. For example: 'www.example.com, example.com'

updateApp_description :: Lens' UpdateApp (Maybe Text)

A description of the app.

updateApp_appId :: Lens' UpdateApp Text

The app ID.

Destructuring the Response

data UpdateAppResponse

See: newUpdateAppResponse smart constructor.

Constructors

UpdateAppResponse'

newUpdateAppResponse :: UpdateAppResponse

Create a value of UpdateAppResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.