Network.AWS.OpsWorks.UpdateInstance

Description

Updates a specified instance.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Creating a Request

data UpdateInstance

See: newUpdateInstance smart constructor.

Constructors

UpdateInstance'

Fields

installUpdatesOnBoot :: Maybe Bool

Whether to install operating system and package updates when the instance boots. The default value is true. To control when updates are installed, set this value to false. You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or by manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

hostname :: Maybe Text

The instance host name.

sshKeyName :: Maybe Text

The instance's Amazon EC2 key name.

agentVersion :: Maybe Text

The default AWS OpsWorks Stacks agent version. You have the following options:

INHERIT - Use the stack's default agent version setting.

version_number - Use the specified agent version. This value overrides the stack's default setting. To update the agent version, you must edit the instance configuration and specify a new version. AWS OpsWorks Stacks then automatically installs that version on the instance.

The default setting is INHERIT. To specify an agent version, you must use the complete version number, not the abbreviated number shown on the console. For a list of available agent version numbers, call DescribeAgentVersions.

AgentVersion cannot be set to Chef 12.2.

instanceType :: Maybe Text

The instance type, such as t2.micro. For a list of supported instance types, open the stack in the console, choose Instances, and choose + Instance. The Size list contains the currently supported types. For more information, see Instance Families and Types. The parameter values that you use to specify the various types are in the API Name column of the Available Instance Types table.

ebsOptimized :: Maybe Bool

This property cannot be updated.

os :: Maybe Text

The instance's operating system, which must be set to one of the following. You cannot update an instance that is using a custom AMI.

A supported Linux operating system: An Amazon Linux version, such as Amazon Linux 2018.03, Amazon Linux 2017.09, Amazon Linux 2017.03, Amazon Linux 2016.09, Amazon Linux 2016.03, Amazon Linux 2015.09, or Amazon Linux 2015.03.

A supported Ubuntu operating system, such as Ubuntu 16.04 LTS, Ubuntu 14.04 LTS, or Ubuntu 12.04 LTS.

CentOS Linux 7

Red Hat Enterprise Linux 7

A supported Windows operating system, such as Microsoft Windows Server 2012 R2 Base, Microsoft Windows Server 2012 R2 with SQL Server Express, Microsoft Windows Server 2012 R2 with SQL Server Standard, or Microsoft Windows Server 2012 R2 with SQL Server Web.

For more information about supported operating systems, see AWS OpsWorks Stacks Operating Systems.

The default option is the current Amazon Linux version. If you set this parameter to Custom, you must use the AmiId parameter to specify the custom AMI that you want to use. For more information about supported operating systems, see Operating Systems. For more information about how to use custom AMIs with OpsWorks, see Using Custom AMIs.

You can specify a different Linux operating system for the updated stack, but you cannot change from Linux to Windows or Windows to Linux.

autoScalingType :: Maybe AutoScalingType

For load-based or time-based instances, the type. Windows stacks can use only time-based instances.

layerIds :: Maybe [Text]

The instance's layer IDs.

architecture :: Maybe Architecture

The instance architecture. Instance types do not necessarily support both architectures. For a list of the architectures that are supported by the different instance types, see Instance Families and Types.

amiId :: Maybe Text

The ID of the AMI that was used to create the instance. The value of this parameter must be the same AMI ID that the instance is already using. You cannot apply a new AMI to an instance by running UpdateInstance. UpdateInstance does not work on instances that are using custom AMIs.

instanceId :: Text

The instance ID.

newUpdateInstance

Arguments

:: Text
$sel:instanceId:UpdateInstance'

-> UpdateInstance

Create a value of UpdateInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:installUpdatesOnBoot:UpdateInstance', updateInstance_installUpdatesOnBoot - Whether to install operating system and package updates when the instance boots. The default value is true. To control when updates are installed, set this value to false. You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or by manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

$sel:hostname:UpdateInstance', updateInstance_hostname - The instance host name.

$sel:sshKeyName:UpdateInstance', updateInstance_sshKeyName - The instance's Amazon EC2 key name.

$sel:agentVersion:UpdateInstance', updateInstance_agentVersion - The default AWS OpsWorks Stacks agent version. You have the following options:

INHERIT - Use the stack's default agent version setting.

version_number - Use the specified agent version. This value overrides the stack's default setting. To update the agent version, you must edit the instance configuration and specify a new version. AWS OpsWorks Stacks then automatically installs that version on the instance.

The default setting is INHERIT. To specify an agent version, you must use the complete version number, not the abbreviated number shown on the console. For a list of available agent version numbers, call DescribeAgentVersions.

AgentVersion cannot be set to Chef 12.2.

$sel:instanceType:UpdateInstance', updateInstance_instanceType - The instance type, such as t2.micro. For a list of supported instance types, open the stack in the console, choose Instances, and choose + Instance. The Size list contains the currently supported types. For more information, see Instance Families and Types. The parameter values that you use to specify the various types are in the API Name column of the Available Instance Types table.

$sel:ebsOptimized:UpdateInstance', updateInstance_ebsOptimized - This property cannot be updated.

$sel:os:UpdateInstance', updateInstance_os - The instance's operating system, which must be set to one of the following. You cannot update an instance that is using a custom AMI.

A supported Linux operating system: An Amazon Linux version, such as Amazon Linux 2018.03, Amazon Linux 2017.09, Amazon Linux 2017.03, Amazon Linux 2016.09, Amazon Linux 2016.03, Amazon Linux 2015.09, or Amazon Linux 2015.03.

A supported Ubuntu operating system, such as Ubuntu 16.04 LTS, Ubuntu 14.04 LTS, or Ubuntu 12.04 LTS.

CentOS Linux 7

Red Hat Enterprise Linux 7

A supported Windows operating system, such as Microsoft Windows Server 2012 R2 Base, Microsoft Windows Server 2012 R2 with SQL Server Express, Microsoft Windows Server 2012 R2 with SQL Server Standard, or Microsoft Windows Server 2012 R2 with SQL Server Web.

For more information about supported operating systems, see AWS OpsWorks Stacks Operating Systems.

The default option is the current Amazon Linux version. If you set this parameter to Custom, you must use the AmiId parameter to specify the custom AMI that you want to use. For more information about supported operating systems, see Operating Systems. For more information about how to use custom AMIs with OpsWorks, see Using Custom AMIs.

You can specify a different Linux operating system for the updated stack, but you cannot change from Linux to Windows or Windows to Linux.

$sel:autoScalingType:UpdateInstance', updateInstance_autoScalingType - For load-based or time-based instances, the type. Windows stacks can use only time-based instances.

$sel:layerIds:UpdateInstance', updateInstance_layerIds - The instance's layer IDs.

$sel:architecture:UpdateInstance', updateInstance_architecture - The instance architecture. Instance types do not necessarily support both architectures. For a list of the architectures that are supported by the different instance types, see Instance Families and Types.

$sel:amiId:UpdateInstance', updateInstance_amiId - The ID of the AMI that was used to create the instance. The value of this parameter must be the same AMI ID that the instance is already using. You cannot apply a new AMI to an instance by running UpdateInstance. UpdateInstance does not work on instances that are using custom AMIs.

$sel:instanceId:UpdateInstance', updateInstance_instanceId - The instance ID.

Request Lenses

updateInstance_installUpdatesOnBoot :: Lens' UpdateInstance (Maybe Bool)

Whether to install operating system and package updates when the instance boots. The default value is true. To control when updates are installed, set this value to false. You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or by manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

updateInstance_hostname :: Lens' UpdateInstance (Maybe Text)

The instance host name.

updateInstance_sshKeyName :: Lens' UpdateInstance (Maybe Text)

The instance's Amazon EC2 key name.

updateInstance_agentVersion :: Lens' UpdateInstance (Maybe Text)

The default AWS OpsWorks Stacks agent version. You have the following options:

INHERIT - Use the stack's default agent version setting.

version_number - Use the specified agent version. This value overrides the stack's default setting. To update the agent version, you must edit the instance configuration and specify a new version. AWS OpsWorks Stacks then automatically installs that version on the instance.

The default setting is INHERIT. To specify an agent version, you must use the complete version number, not the abbreviated number shown on the console. For a list of available agent version numbers, call DescribeAgentVersions.

AgentVersion cannot be set to Chef 12.2.

updateInstance_instanceType :: Lens' UpdateInstance (Maybe Text)

The instance type, such as t2.micro. For a list of supported instance types, open the stack in the console, choose Instances, and choose + Instance. The Size list contains the currently supported types. For more information, see Instance Families and Types. The parameter values that you use to specify the various types are in the API Name column of the Available Instance Types table.

updateInstance_ebsOptimized :: Lens' UpdateInstance (Maybe Bool)

This property cannot be updated.

updateInstance_os :: Lens' UpdateInstance (Maybe Text)

The instance's operating system, which must be set to one of the following. You cannot update an instance that is using a custom AMI.

A supported Linux operating system: An Amazon Linux version, such as Amazon Linux 2018.03, Amazon Linux 2017.09, Amazon Linux 2017.03, Amazon Linux 2016.09, Amazon Linux 2016.03, Amazon Linux 2015.09, or Amazon Linux 2015.03.

A supported Ubuntu operating system, such as Ubuntu 16.04 LTS, Ubuntu 14.04 LTS, or Ubuntu 12.04 LTS.

CentOS Linux 7

Red Hat Enterprise Linux 7

A supported Windows operating system, such as Microsoft Windows Server 2012 R2 Base, Microsoft Windows Server 2012 R2 with SQL Server Express, Microsoft Windows Server 2012 R2 with SQL Server Standard, or Microsoft Windows Server 2012 R2 with SQL Server Web.

For more information about supported operating systems, see AWS OpsWorks Stacks Operating Systems.

The default option is the current Amazon Linux version. If you set this parameter to Custom, you must use the AmiId parameter to specify the custom AMI that you want to use. For more information about supported operating systems, see Operating Systems. For more information about how to use custom AMIs with OpsWorks, see Using Custom AMIs.

You can specify a different Linux operating system for the updated stack, but you cannot change from Linux to Windows or Windows to Linux.

updateInstance_autoScalingType :: Lens' UpdateInstance (Maybe AutoScalingType)

For load-based or time-based instances, the type. Windows stacks can use only time-based instances.

updateInstance_layerIds :: Lens' UpdateInstance (Maybe [Text])

The instance's layer IDs.

updateInstance_architecture :: Lens' UpdateInstance (Maybe Architecture)

The instance architecture. Instance types do not necessarily support both architectures. For a list of the architectures that are supported by the different instance types, see Instance Families and Types.

updateInstance_amiId :: Lens' UpdateInstance (Maybe Text)

The ID of the AMI that was used to create the instance. The value of this parameter must be the same AMI ID that the instance is already using. You cannot apply a new AMI to an instance by running UpdateInstance. UpdateInstance does not work on instances that are using custom AMIs.

updateInstance_instanceId :: Lens' UpdateInstance Text

The instance ID.

Destructuring the Response

data UpdateInstanceResponse

See: newUpdateInstanceResponse smart constructor.

Constructors

UpdateInstanceResponse'

newUpdateInstanceResponse :: UpdateInstanceResponse

Create a value of UpdateInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.