Network.AWS.OpsWorks.UpdateLayer

Description

Updates a specified layer.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Creating a Request

data UpdateLayer

See: newUpdateLayer smart constructor.

Constructors

UpdateLayer'

Fields

customInstanceProfileArn :: Maybe Text

The ARN of an IAM profile to be used for all of the layer's EC2 instances. For more information about IAM ARNs, see Using Identifiers.

customSecurityGroupIds :: Maybe [Text]

An array containing the layer's custom security group IDs.

installUpdatesOnBoot :: Maybe Bool

Whether to install operating system and package updates when the instance boots. The default value is true. To control when updates are installed, set this value to false. You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

cloudWatchLogsConfiguration :: Maybe CloudWatchLogsConfiguration

Specifies CloudWatch Logs configuration options for the layer. For more information, see CloudWatchLogsLogStream.

lifecycleEventConfiguration :: Maybe LifecycleEventConfiguration

shortname :: Maybe Text

For custom layers only, use this parameter to specify the layer's short name, which is used internally by AWS OpsWorks Stacks and by Chef. The short name is also used as the name for the directory where your app files are installed. It can have a maximum of 200 characters and must be in the following format: /\A[a-z0-9\-\_\.]+\Z/.

The built-in layers' short names are defined by AWS OpsWorks Stacks. For more information, see the Layer Reference

customRecipes :: Maybe Recipes

A LayerCustomRecipes object that specifies the layer's custom recipes.

customJson :: Maybe Text

A JSON-formatted string containing custom stack configuration and deployment attributes to be installed on the layer's instances. For more information, see Using Custom JSON.

volumeConfigurations :: Maybe [VolumeConfiguration]

A VolumeConfigurations object that describes the layer's Amazon EBS volumes.

enableAutoHealing :: Maybe Bool

Whether to disable auto healing for the layer.

packages :: Maybe [Text]

An array of Package objects that describe the layer's packages.

attributes :: Maybe (HashMap LayerAttributesKeys (Maybe Text))

One or more user-defined key/value pairs to be added to the stack attributes.

name :: Maybe Text

The layer name, which is used by the console.

autoAssignPublicIps :: Maybe Bool

For stacks that are running in a VPC, whether to automatically assign a public IP address to the layer's instances. For more information, see How to Edit a Layer.

useEbsOptimizedInstances :: Maybe Bool

Whether to use Amazon EBS-optimized instances.

autoAssignElasticIps :: Maybe Bool

Whether to automatically assign an Elastic IP address to the layer's instances. For more information, see How to Edit a Layer.

layerId :: Text

The layer ID.

newUpdateLayer

Arguments

:: Text
$sel:layerId:UpdateLayer'

-> UpdateLayer

Create a value of UpdateLayer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customInstanceProfileArn:UpdateLayer', updateLayer_customInstanceProfileArn - The ARN of an IAM profile to be used for all of the layer's EC2 instances. For more information about IAM ARNs, see Using Identifiers.

$sel:customSecurityGroupIds:UpdateLayer', updateLayer_customSecurityGroupIds - An array containing the layer's custom security group IDs.

$sel:installUpdatesOnBoot:UpdateLayer', updateLayer_installUpdatesOnBoot - Whether to install operating system and package updates when the instance boots. The default value is true. To control when updates are installed, set this value to false. You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

$sel:cloudWatchLogsConfiguration:UpdateLayer', updateLayer_cloudWatchLogsConfiguration - Specifies CloudWatch Logs configuration options for the layer. For more information, see CloudWatchLogsLogStream.

$sel:lifecycleEventConfiguration:UpdateLayer', updateLayer_lifecycleEventConfiguration -

$sel:shortname:UpdateLayer', updateLayer_shortname - For custom layers only, use this parameter to specify the layer's short name, which is used internally by AWS OpsWorks Stacks and by Chef. The short name is also used as the name for the directory where your app files are installed. It can have a maximum of 200 characters and must be in the following format: /\A[a-z0-9\-\_\.]+\Z/.

The built-in layers' short names are defined by AWS OpsWorks Stacks. For more information, see the Layer Reference

$sel:customRecipes:UpdateLayer', updateLayer_customRecipes - A LayerCustomRecipes object that specifies the layer's custom recipes.

$sel:customJson:UpdateLayer', updateLayer_customJson - A JSON-formatted string containing custom stack configuration and deployment attributes to be installed on the layer's instances. For more information, see Using Custom JSON.

$sel:volumeConfigurations:UpdateLayer', updateLayer_volumeConfigurations - A VolumeConfigurations object that describes the layer's Amazon EBS volumes.

$sel:enableAutoHealing:UpdateLayer', updateLayer_enableAutoHealing - Whether to disable auto healing for the layer.

$sel:packages:UpdateLayer', updateLayer_packages - An array of Package objects that describe the layer's packages.

$sel:attributes:UpdateLayer', updateLayer_attributes - One or more user-defined key/value pairs to be added to the stack attributes.

$sel:name:UpdateLayer', updateLayer_name - The layer name, which is used by the console.

$sel:autoAssignPublicIps:UpdateLayer', updateLayer_autoAssignPublicIps - For stacks that are running in a VPC, whether to automatically assign a public IP address to the layer's instances. For more information, see How to Edit a Layer.

$sel:useEbsOptimizedInstances:UpdateLayer', updateLayer_useEbsOptimizedInstances - Whether to use Amazon EBS-optimized instances.

$sel:autoAssignElasticIps:UpdateLayer', updateLayer_autoAssignElasticIps - Whether to automatically assign an Elastic IP address to the layer's instances. For more information, see How to Edit a Layer.

$sel:layerId:UpdateLayer', updateLayer_layerId - The layer ID.

Request Lenses

updateLayer_customInstanceProfileArn :: Lens' UpdateLayer (Maybe Text)

The ARN of an IAM profile to be used for all of the layer's EC2 instances. For more information about IAM ARNs, see Using Identifiers.

updateLayer_customSecurityGroupIds :: Lens' UpdateLayer (Maybe [Text])

An array containing the layer's custom security group IDs.

updateLayer_installUpdatesOnBoot :: Lens' UpdateLayer (Maybe Bool)

Whether to install operating system and package updates when the instance boots. The default value is true. To control when updates are installed, set this value to false. You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances.

We strongly recommend using the default value of true, to ensure that your instances have the latest security updates.

updateLayer_cloudWatchLogsConfiguration :: Lens' UpdateLayer (Maybe CloudWatchLogsConfiguration)

Specifies CloudWatch Logs configuration options for the layer. For more information, see CloudWatchLogsLogStream.

updateLayer_lifecycleEventConfiguration :: Lens' UpdateLayer (Maybe LifecycleEventConfiguration)

updateLayer_shortname :: Lens' UpdateLayer (Maybe Text)

For custom layers only, use this parameter to specify the layer's short name, which is used internally by AWS OpsWorks Stacks and by Chef. The short name is also used as the name for the directory where your app files are installed. It can have a maximum of 200 characters and must be in the following format: /\A[a-z0-9\-\_\.]+\Z/.

The built-in layers' short names are defined by AWS OpsWorks Stacks. For more information, see the Layer Reference

updateLayer_customRecipes :: Lens' UpdateLayer (Maybe Recipes)

A LayerCustomRecipes object that specifies the layer's custom recipes.

updateLayer_customJson :: Lens' UpdateLayer (Maybe Text)

A JSON-formatted string containing custom stack configuration and deployment attributes to be installed on the layer's instances. For more information, see Using Custom JSON.

updateLayer_volumeConfigurations :: Lens' UpdateLayer (Maybe [VolumeConfiguration])

A VolumeConfigurations object that describes the layer's Amazon EBS volumes.

updateLayer_enableAutoHealing :: Lens' UpdateLayer (Maybe Bool)

Whether to disable auto healing for the layer.

updateLayer_packages :: Lens' UpdateLayer (Maybe [Text])

An array of Package objects that describe the layer's packages.

updateLayer_attributes :: Lens' UpdateLayer (Maybe (HashMap LayerAttributesKeys (Maybe Text)))

One or more user-defined key/value pairs to be added to the stack attributes.

updateLayer_name :: Lens' UpdateLayer (Maybe Text)

The layer name, which is used by the console.

updateLayer_autoAssignPublicIps :: Lens' UpdateLayer (Maybe Bool)

For stacks that are running in a VPC, whether to automatically assign a public IP address to the layer's instances. For more information, see How to Edit a Layer.

updateLayer_useEbsOptimizedInstances :: Lens' UpdateLayer (Maybe Bool)

Whether to use Amazon EBS-optimized instances.

updateLayer_autoAssignElasticIps :: Lens' UpdateLayer (Maybe Bool)

Whether to automatically assign an Elastic IP address to the layer's instances. For more information, see How to Edit a Layer.

updateLayer_layerId :: Lens' UpdateLayer Text

The layer ID.

Destructuring the Response

data UpdateLayerResponse

See: newUpdateLayerResponse smart constructor.

Constructors

UpdateLayerResponse'

newUpdateLayerResponse :: UpdateLayerResponse

Create a value of UpdateLayerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.