Network.AWS.OpsWorks.UpdateVolume

Description

Updates an Amazon EBS volume's name or mount point. For more information, see Resource Management.

Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Creating a Request

data UpdateVolume

See: newUpdateVolume smart constructor.

Constructors

UpdateVolume'

Fields

name :: Maybe Text

The new name.

mountPoint :: Maybe Text

The new mount point.

volumeId :: Text

The volume ID.

newUpdateVolume

Arguments

:: Text
$sel:volumeId:UpdateVolume'

-> UpdateVolume

Create a value of UpdateVolume with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:UpdateVolume', updateVolume_name - The new name.

$sel:mountPoint:UpdateVolume', updateVolume_mountPoint - The new mount point.

$sel:volumeId:UpdateVolume', updateVolume_volumeId - The volume ID.

Request Lenses

updateVolume_name :: Lens' UpdateVolume (Maybe Text)

The new name.

updateVolume_mountPoint :: Lens' UpdateVolume (Maybe Text)

The new mount point.

updateVolume_volumeId :: Lens' UpdateVolume Text

The volume ID.

Destructuring the Response

data UpdateVolumeResponse

See: newUpdateVolumeResponse smart constructor.

Constructors

UpdateVolumeResponse'

newUpdateVolumeResponse :: UpdateVolumeResponse

Create a value of UpdateVolumeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.