Network.AWS.OpsWorks.V2013_02_18.CreateInstance

Description

Creates an instance in a specified stack. For more information, see Adding an Instance to a Layer. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

createInstance

Arguments

:: Text
_cirInstanceType

-> Text
_cirStackId

-> [Text]
_cirLayerIds

-> CreateInstance

Minimum specification for a CreateInstance request.

data CreateInstance

Constructors

CreateInstance

Fields

_cirInstanceType :: Text

The instance type. AWS OpsWorks supports all instance types except Cluster Compute, Cluster GPU, and High Memory Cluster. For more information, see Instance Families and Types. The parameter values that you use to specify the various types are in the API Name column of the Available Instance Types table.

_cirStackId :: Text

The stack ID.

_cirLayerIds :: [Text]

An array that contains the instance layer IDs.

_cirArchitecture :: Maybe Architecture

The instance architecture. The default option is x86_64. Instance types do not necessarily support both architectures. For a list of the architectures that are supported by the different instance types, see Instance Families and Types.

_cirAutoScalingType :: Maybe AutoScalingType

The instance auto scaling type, which has three possible values: AlwaysRunning: A 24/7 instance, which is not affected by auto scaling. TimeBasedAutoScaling: A time-based auto scaling instance, which is started and stopped based on a specified schedule. To specify the schedule, call SetTimeBasedAutoScaling. LoadBasedAutoScaling: A load-based auto scaling instance, which is started and stopped based on load metrics. To use load-based auto scaling, you must enable it for the instance layer and configure the thresholds by calling SetLoadBasedAutoScaling.

_cirInstallUpdatesOnBoot :: Maybe Bool

Whether to install operating system and package updates when the instance boots. The default value is true. To control when updates are installed, set this value to false. You must then update your instances manually by using CreateDeployment to run the update_dependencies stack command or manually running yum (Amazon Linux) or apt-get (Ubuntu) on the instances. We strongly recommend using the default value of true to ensure that your instances have the latest security updates.

_cirEbsOptimized :: Maybe Bool

Whether to create an Amazon EBS-optimized instance.

_cirRootDeviceType :: Maybe RootDeviceType

The instance root device type. For more information, see Storage for the Root Device.

_cirVirtualizationType :: Maybe Text

The instance's virtualization type, paravirtual or hvm.

_cirHostname :: Maybe Text

The instance host name.

_cirSshKeyName :: Maybe Text

The instance SSH key name.

_cirSubnetId :: Maybe Text

The ID of the instance's subnet. If the stack is running in a VPC, you can use this parameter to override the stack's default subnet ID value and direct AWS OpsWorks to launch the instance in a different subnet.

_cirOs :: Maybe Text

The instance operating system, which must be set to one of the following. Standard operating systems: Amazon Linux or Ubuntu 12.04 LTS Custom AMIs: Custom The default option is Amazon Linux. If you set this parameter to Custom, you must use the CreateInstance action's AmiId parameter to specify the custom AMI that you want to use. For more information on the standard operating systems, see Operating SystemsFor more information on how to use custom AMIs with OpsWorks, see Using Custom AMIs.

_cirAvailabilityZone :: Maybe Text

The instance Availability Zone. For more information, see Regions and Endpoints.

_cirAmiId :: Maybe Text

A custom AMI ID to be used to create the instance. The AMI should be based on one of the standard AWS OpsWorks APIs: Amazon Linux or Ubuntu 12.04 LTS. For more information, see Instances.

cirVirtualizationType :: Lens' CreateInstance (Maybe Text)

cirSubnetId :: Lens' CreateInstance (Maybe Text)

cirStackId :: Lens' CreateInstance Text

cirSshKeyName :: Lens' CreateInstance (Maybe Text)

cirRootDeviceType :: Lens' CreateInstance (Maybe RootDeviceType)

cirOs :: Lens' CreateInstance (Maybe Text)

cirLayerIds :: Lens' CreateInstance [Text]

cirInstanceType :: Lens' CreateInstance Text

cirInstallUpdatesOnBoot :: Lens' CreateInstance (Maybe Bool)

cirHostname :: Lens' CreateInstance (Maybe Text)

cirEbsOptimized :: Lens' CreateInstance (Maybe Bool)

cirAvailabilityZone :: Lens' CreateInstance (Maybe Text)

cirAutoScalingType :: Lens' CreateInstance (Maybe AutoScalingType)

cirArchitecture :: Lens' CreateInstance (Maybe Architecture)

cirAmiId :: Lens' CreateInstance (Maybe Text)

data CreateInstanceResponse

Constructors

CreateInstanceResponse

Fields

_cisInstanceId :: Maybe Text

The instance ID.

cisInstanceId :: Iso' CreateInstanceResponse (Maybe Text)