Network.AWS.OpsWorks.V2013_02_18.DeleteUserProfile

Description

Deletes a user profile. Required Permissions: To use this action, an IAM user must have an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

data DeleteUserProfile

Constructors

DeleteUserProfile

Fields

_duptIamUserArn :: Text

The user's IAM ARN.

duptIamUserArn :: Iso' DeleteUserProfile Text

data DeleteUserProfileResponse

Constructors

DeleteUserProfileResponse