Network.AWS.OpsWorks.V2013_02_18.DeregisterVolume

Description

Deregisters an Amazon EBS volume. The volume can then be registered by another stack. For more information, see Resource Management. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

data DeregisterVolume

Constructors

DeregisterVolume

Fields

_dvrVolumeId :: Text

The volume ID.

dvrVolumeId :: Iso' DeregisterVolume Text

data DeregisterVolumeResponse

Constructors

DeregisterVolumeResponse