Network.AWS.OpsWorks.V2013_02_18.DescribeCommands

Description

Describes the results of specified commands. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

describeCommands :: DescribeCommands

Minimum specification for a DescribeCommands request.

data DescribeCommands

Constructors

DescribeCommands

Fields

_dcrDeploymentId :: Maybe Text

The deployment ID. If you include this parameter, DescribeCommands returns a description of the commands associated with the specified deployment.

_dcrInstanceId :: Maybe Text

The instance ID. If you include this parameter, DescribeCommands returns a description of the commands associated with the specified instance.

_dcrCommandIds :: [Text]

An array of command IDs. If you include this parameter, DescribeCommands returns a description of the specified commands. Otherwise, it returns a description of every command.

dcrInstanceId :: Lens' DescribeCommands (Maybe Text)

dcrDeploymentId :: Lens' DescribeCommands (Maybe Text)

dcrCommandIds :: Lens' DescribeCommands [Text]

data DescribeCommandsResponse

Constructors

DescribeCommandsResponse

Fields

_dcsCommands :: [Command]

An array of Command objects that describe each of the specified commands.

dcsCommands :: Iso' DescribeCommandsResponse [Command]