Network.AWS.OpsWorks.V2013_02_18.DescribeRaidArrays

Description

Describe an instance's RAID arrays. You must specify at least one of the parameters. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

describeRaidArrays :: DescribeRaidArrays

Minimum specification for a DescribeRaidArrays request.

data DescribeRaidArrays

Constructors

DescribeRaidArrays

Fields

_drarInstanceId :: Maybe Text

The instance ID. If you use this parameter, DescribeRaidArrays returns descriptions of the RAID arrays associated with the specified instance.

_drarRaidArrayIds :: [Text]

An array of RAID array IDs. If you use this parameter, DescribeRaidArrays returns descriptions of the specified arrays. Otherwise, it returns a description of every array.

drarRaidArrayIds :: Lens' DescribeRaidArrays [Text]

drarInstanceId :: Lens' DescribeRaidArrays (Maybe Text)

data DescribeRaidArraysResponse

Constructors

DescribeRaidArraysResponse

Fields

_drasRaidArrays :: [RaidArray]

A RaidArrays object that describes the specified RAID arrays.

drasRaidArrays :: Iso' DescribeRaidArraysResponse [RaidArray]