Network.AWS.OpsWorks.V2013_02_18.DescribeStackSummary

Description

Describes the number of layers and apps in a specified stack, and the number of instances in each state, such as running_setup or online. Required Permissions: To use this action, an IAM user must have a Show, Deploy, or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

data DescribeStackSummary

Constructors

DescribeStackSummary

Fields

_dssrStackId :: Text

The stack ID.

dssrStackId :: Iso' DescribeStackSummary Text

data DescribeStackSummaryResponse

Constructors

DescribeStackSummaryResponse

Fields

_dsssStackSummary :: Maybe StackSummary

A StackSummary object that contains the results.

dsssStackSummary :: Iso' DescribeStackSummaryResponse (Maybe StackSummary)