Network.AWS.OpsWorks.V2013_02_18.DetachElasticLoadBalancer

Description

Detaches a specified Elastic Load Balancing instance from its layer. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

data DetachElasticLoadBalancer

Constructors

DetachElasticLoadBalancer

Fields

_delbrElasticLoadBalancerName :: Text

The Elastic Load Balancing instance's name.

_delbrLayerId :: Text

The ID of the layer that the Elastic Load Balancing instance is attached to.

delbrLayerId :: Lens' DetachElasticLoadBalancer Text

delbrElasticLoadBalancerName :: Lens' DetachElasticLoadBalancer Text

data DetachElasticLoadBalancerResponse

Constructors

DetachElasticLoadBalancerResponse