Network.AWS.OpsWorks.V2013_02_18.SetTimeBasedAutoScaling

Description

Specify the time-based auto scaling configuration for a specified instance. For more information, see Managing Load with Time-based and Load-based Instances. Required Permissions: To use this action, an IAM user must have a Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

setTimeBasedAutoScaling

Arguments

:: Text
_stbasrInstanceId

-> SetTimeBasedAutoScaling

Minimum specification for a SetTimeBasedAutoScaling request.

data SetTimeBasedAutoScaling

Constructors

SetTimeBasedAutoScaling

Fields

_stbasrInstanceId :: Text

The instance ID.

_stbasrAutoScalingSchedule :: Maybe WeeklyAutoScalingSchedule

An AutoScalingSchedule with the instance schedule.

stbasrInstanceId :: Lens' SetTimeBasedAutoScaling Text

stbasrAutoScalingSchedule :: Lens' SetTimeBasedAutoScaling (Maybe WeeklyAutoScalingSchedule)

data SetTimeBasedAutoScalingResponse

Constructors

SetTimeBasedAutoScalingResponse