Network.AWS.OpsWorks.V2013_02_18.UpdateApp

Description

Updates a specified app. Required Permissions: To use this action, an IAM user must have a Deploy or Manage permissions level for the stack, or an attached policy that explicitly grants permissions. For more information on user permissions, see Managing User Permissions.

Documentation

updateApp

Arguments

:: Text
_uarAppId

-> UpdateApp

Minimum specification for a UpdateApp request.

data UpdateApp

Constructors

UpdateApp

Fields

_uarAppId :: Text

The app ID.

_uarAttributes :: HashMap AppAttributesKeys Text

One or more user-defined key/value pairs to be added to the stack attributes.

_uarType :: Maybe AppType

The app type.

_uarEnableSsl :: Maybe Bool

Whether SSL is enabled for the app.

_uarDataSources :: [DataSource]

The app's data sources.

_uarAppSource :: Maybe Source

A Source object that specifies the app repository.

_uarSslConfiguration :: Maybe SslConfiguration

An SslConfiguration object with the SSL configuration.

_uarName :: Maybe Text

The app name.

_uarDescription :: Maybe Text

A description of the app.

_uarDomains :: [Text]

The app's virtual host settings, with multiple domains separated by commas. For example: 'www.example.com, example.com'.

uarType :: Lens' UpdateApp (Maybe AppType)

uarSslConfiguration :: Lens' UpdateApp (Maybe SslConfiguration)

uarName :: Lens' UpdateApp (Maybe Text)

uarEnableSsl :: Lens' UpdateApp (Maybe Bool)

uarDomains :: Lens' UpdateApp [Text]

uarDescription :: Lens' UpdateApp (Maybe Text)

uarDataSources :: Lens' UpdateApp [DataSource]

uarAttributes :: Lens' UpdateApp (HashMap AppAttributesKeys Text)

uarAppSource :: Lens' UpdateApp (Maybe Source)

uarAppId :: Lens' UpdateApp Text

data UpdateAppResponse

Constructors

UpdateAppResponse