Network.AWS.OpsWorksCM.AssociateNode

Description

Associates a new node with the server. For more information about how to disassociate a node, see DisassociateNode.

On a Chef server: This command is an alternative to knife bootstrap.

Example (Chef): aws opsworks-cm associate-node --server-name MyServer --node-name MyManagedNode --engine-attributes "Name=CHEF_ORGANIZATION,Value=default" "Name=CHEF_NODE_PUBLIC_KEY,Value=public-key-pem"

On a Puppet server, this command is an alternative to the puppet cert sign command that signs a Puppet node CSR.

Example (Puppet): aws opsworks-cm associate-node --server-name MyServer --node-name MyManagedNode --engine-attributes "Name=PUPPET_NODE_CSR,Value=csr-pem"

A node can can only be associated with servers that are in a HEALTHY state. Otherwise, an InvalidStateException is thrown. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid. The AssociateNode API call can be integrated into Auto Scaling configurations, AWS Cloudformation templates, or the user data of a server's instance.

Creating a Request

data AssociateNode

See: newAssociateNode smart constructor.

Constructors

AssociateNode'

Fields

serverName :: Text

The name of the server with which to associate the node.

nodeName :: Text

The name of the node.

engineAttributes :: [EngineAttribute]

Engine attributes used for associating the node.

Attributes accepted in a AssociateNode request for Chef

CHEF_ORGANIZATION: The Chef organization with which the node is associated. By default only one organization named default can exist.

CHEF_NODE_PUBLIC_KEY: A PEM-formatted public key. This key is required for the chef-client agent to access the Chef API.

Attributes accepted in a AssociateNode request for Puppet

PUPPET_NODE_CSR: A PEM-formatted certificate-signing request (CSR) that is created by the node.

newAssociateNode

Arguments

:: Text
$sel:serverName:AssociateNode'

-> Text
$sel:nodeName:AssociateNode'

-> AssociateNode

Create a value of AssociateNode with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serverName:AssociateNode', associateNode_serverName - The name of the server with which to associate the node.

$sel:nodeName:AssociateNode', associateNode_nodeName - The name of the node.

$sel:engineAttributes:AssociateNode', associateNode_engineAttributes - Engine attributes used for associating the node.

Attributes accepted in a AssociateNode request for Chef

CHEF_ORGANIZATION: The Chef organization with which the node is associated. By default only one organization named default can exist.

CHEF_NODE_PUBLIC_KEY: A PEM-formatted public key. This key is required for the chef-client agent to access the Chef API.

Attributes accepted in a AssociateNode request for Puppet

PUPPET_NODE_CSR: A PEM-formatted certificate-signing request (CSR) that is created by the node.

Request Lenses

associateNode_serverName :: Lens' AssociateNode Text

The name of the server with which to associate the node.

associateNode_nodeName :: Lens' AssociateNode Text

The name of the node.

associateNode_engineAttributes :: Lens' AssociateNode [EngineAttribute]

Engine attributes used for associating the node.

Attributes accepted in a AssociateNode request for Chef

CHEF_ORGANIZATION: The Chef organization with which the node is associated. By default only one organization named default can exist.

CHEF_NODE_PUBLIC_KEY: A PEM-formatted public key. This key is required for the chef-client agent to access the Chef API.

Attributes accepted in a AssociateNode request for Puppet

PUPPET_NODE_CSR: A PEM-formatted certificate-signing request (CSR) that is created by the node.

Destructuring the Response

data AssociateNodeResponse

See: newAssociateNodeResponse smart constructor.

Constructors

AssociateNodeResponse'

Fields

nodeAssociationStatusToken :: Maybe Text

Contains a token which can be passed to the DescribeNodeAssociationStatus API call to get the status of the association request.

httpStatus :: Int

The response's http status code.

newAssociateNodeResponse

Arguments

:: Int
$sel:httpStatus:AssociateNodeResponse'

-> AssociateNodeResponse

Create a value of AssociateNodeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nodeAssociationStatusToken:AssociateNodeResponse', associateNodeResponse_nodeAssociationStatusToken - Contains a token which can be passed to the DescribeNodeAssociationStatus API call to get the status of the association request.

$sel:httpStatus:AssociateNodeResponse', associateNodeResponse_httpStatus - The response's http status code.

Response Lenses

associateNodeResponse_nodeAssociationStatusToken :: Lens' AssociateNodeResponse (Maybe Text)

Contains a token which can be passed to the DescribeNodeAssociationStatus API call to get the status of the association request.

associateNodeResponse_httpStatus :: Lens' AssociateNodeResponse Int

The response's http status code.