Network.AWS.OpsWorksCM.CreateServer

Description

Creates and immedately starts a new server. The server is ready to use when it is in the HEALTHY state. By default, you can create a maximum of 10 servers.

This operation is asynchronous.

A LimitExceededException is thrown when you have created the maximum number of servers (10). A ResourceAlreadyExistsException is thrown when a server with the same name already exists in the account. A ResourceNotFoundException is thrown when you specify a backup ID that is not valid or is for a backup that does not exist. A ValidationException is thrown when parameters of the request are not valid.

If you do not specify a security group by adding the SecurityGroupIds parameter, AWS OpsWorks creates a new security group.

Chef Automate: The default security group opens the Chef server to the world on TCP port 443. If a KeyName is present, AWS OpsWorks enables SSH access. SSH is also open to the world on TCP port 22.

Puppet Enterprise: The default security group opens TCP ports 22, 443, 4433, 8140, 8142, 8143, and 8170. If a KeyName is present, AWS OpsWorks enables SSH access. SSH is also open to the world on TCP port 22.

By default, your server is accessible from any IP address. We recommend that you update your security group rules to allow access from known IP addresses and address ranges only. To edit security group rules, open Security Groups in the navigation pane of the EC2 management console.

To specify your own domain for a server, and provide your own self-signed or CA-signed certificate and private key, specify values for CustomDomain, CustomCertificate, and CustomPrivateKey.

Creating a Request

data CreateServer

See: newCreateServer smart constructor.

Constructors

CreateServer'

Fields

preferredBackupWindow :: Maybe Text

The start time for a one-hour period during which AWS OpsWorks CM backs up application-level data on your server if automated backups are enabled. Valid values must be specified in one of the following formats:

HH:MM for daily backups

DDD:HH:MM for weekly backups

MM must be specified as 00. The specified time is in coordinated universal time (UTC). The default value is a random, daily start time.

Example: 08:00, which represents a daily start time of 08:00 UTC.

Example: Mon:08:00, which represents a start time of every Monday at 08:00 UTC. (8:00 a.m.)

securityGroupIds :: Maybe [Text]

A list of security group IDs to attach to the Amazon EC2 instance. If you add this parameter, the specified security groups must be within the VPC that is specified by SubnetIds.

If you do not specify this parameter, AWS OpsWorks CM creates one new security group that uses TCP ports 22 and 443, open to 0.0.0.0/0 (everyone).

disableAutomatedBackup :: Maybe Bool

Enable or disable scheduled backups. Valid values are true or false. The default value is true.

customPrivateKey :: Maybe (Sensitive Text)

A private key in PEM format for connecting to the server by using HTTPS. The private key must not be encrypted; it cannot be protected by a password or passphrase. If you specify a custom private key, you must also specify values for CustomDomain and CustomCertificate.

engineAttributes :: Maybe [EngineAttribute]

Optional engine attributes on a specified server.

Attributes accepted in a Chef createServer request:

CHEF_AUTOMATE_PIVOTAL_KEY: A base64-encoded RSA public key. The corresponding private key is required to access the Chef API. When no CHEF_AUTOMATE_PIVOTAL_KEY is set, a private key is generated and returned in the response.

CHEF_AUTOMATE_ADMIN_PASSWORD: The password for the administrative user in the Chef Automate web-based dashboard. The password length is a minimum of eight characters, and a maximum of 32. The password can contain letters, numbers, and special characters (!/@#$%^&+=_). The password must contain at least one lower case letter, one upper case letter, one number, and one special character. When no CHEF_AUTOMATE_ADMIN_PASSWORD is set, one is generated and returned in the response.

Attributes accepted in a Puppet createServer request:

PUPPET_ADMIN_PASSWORD: To work with the Puppet Enterprise console, a password must use ASCII characters.

PUPPET_R10K_REMOTE: The r10k remote is the URL of your control repository (for example, ssh://git@your.git-repo.com:user/control-repo.git). Specifying an r10k remote opens TCP port 8170.

PUPPET_R10K_PRIVATE_KEY: If you are using a private Git repository, add PUPPET_R10K_PRIVATE_KEY to specify a PEM-encoded private SSH key.

customDomain :: Maybe Text

An optional public endpoint of a server, such as https://aws.my-company.com. To access the server, create a CNAME DNS record in your preferred DNS service that points the custom domain to the endpoint that is generated when the server is created (the value of the CreateServer Endpoint attribute). You cannot access the server by using the generated Endpoint value if the server is using a custom domain. If you specify a custom domain, you must also specify values for CustomCertificate and CustomPrivateKey.

backupId :: Maybe Text

If you specify this field, AWS OpsWorks CM creates the server by using the backup represented by BackupId.

subnetIds :: Maybe [Text]

The IDs of subnets in which to launch the server EC2 instance.

Amazon EC2-Classic customers: This field is required. All servers must run within a VPC. The VPC must have "Auto Assign Public IP" enabled.

EC2-VPC customers: This field is optional. If you do not specify subnet IDs, your EC2 instances are created in a default subnet that is selected by Amazon EC2. If you specify subnet IDs, the VPC must have "Auto Assign Public IP" enabled.

For more information about supported Amazon EC2 platforms, see Supported Platforms.

keyPair :: Maybe Text

The Amazon EC2 key pair to set for the instance. This parameter is optional; if desired, you may specify this parameter to connect to your instances by using SSH.

associatePublicIpAddress :: Maybe Bool

Associate a public IP address with a server that you are launching. Valid values are true or false. The default value is true.

engineVersion :: Maybe Text

The major release version of the engine that you want to use. For a Chef server, the valid value for EngineVersion is currently 2. For a Puppet server, valid values are 2019 or 2017.

preferredMaintenanceWindow :: Maybe Text

The start time for a one-hour period each week during which AWS OpsWorks CM performs maintenance on the instance. Valid values must be specified in the following format: DDD:HH:MM. MM must be specified as 00. The specified time is in coordinated universal time (UTC). The default value is a random one-hour period on Tuesday, Wednesday, or Friday. See TimeWindowDefinition for more information.

Example: Mon:08:00, which represents a start time of every Monday at 08:00 UTC. (8:00 a.m.)

tags :: Maybe [Tag]

A map that contains tag keys and tag values to attach to an AWS OpsWorks for Chef Automate or AWS OpsWorks for Puppet Enterprise server.

The key cannot be empty.

The key can be a maximum of 127 characters, and can contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : / @

The value can be a maximum 255 characters, and contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : / @

Leading and trailing white spaces are trimmed from both the key and value.

A maximum of 50 user-applied tags is allowed for any AWS OpsWorks-CM server.

engineModel :: Maybe Text

The engine model of the server. Valid values in this release include Monolithic for Puppet and Single for Chef.

backupRetentionCount :: Maybe Natural

The number of automated backups that you want to keep. Whenever a new backup is created, AWS OpsWorks CM deletes the oldest backups if this number is exceeded. The default value is 1.

customCertificate :: Maybe Text

A PEM-formatted HTTPS certificate. The value can be be a single, self-signed certificate, or a certificate chain. If you specify a custom certificate, you must also specify values for CustomDomain and CustomPrivateKey. The following are requirements for the CustomCertificate value:

You can provide either a self-signed, custom certificate, or the full certificate chain.

The certificate must be a valid X509 certificate, or a certificate chain in PEM format.

The certificate must be valid at the time of upload. A certificate can't be used before its validity period begins (the certificate's NotBefore date), or after it expires (the certificate's NotAfter date).

The certificate’s common name or subject alternative names (SANs), if present, must match the value of CustomDomain.

The certificate must match the value of CustomPrivateKey.

engine :: Text

The configuration management engine to use. Valid values include ChefAutomate and Puppet.

serverName :: Text

The name of the server. The server name must be unique within your AWS account, within each region. Server names must start with a letter; then letters, numbers, or hyphens (-) are allowed, up to a maximum of 40 characters.

instanceProfileArn :: Text

The ARN of the instance profile that your Amazon EC2 instances use. Although the AWS OpsWorks console typically creates the instance profile for you, if you are using API commands instead, run the service-role-creation.yaml AWS CloudFormation template, located at https://s3.amazonaws.com/opsworks-cm-us-east-1-prod-default-assets/misc/opsworks-cm-roles.yaml. This template creates a CloudFormation stack that includes the instance profile you need.

instanceType :: Text

The Amazon EC2 instance type to use. For example, m5.large.

serviceRoleArn :: Text

The service role that the AWS OpsWorks CM service backend uses to work with your account. Although the AWS OpsWorks management console typically creates the service role for you, if you are using the AWS CLI or API commands, run the service-role-creation.yaml AWS CloudFormation template, located at https://s3.amazonaws.com/opsworks-cm-us-east-1-prod-default-assets/misc/opsworks-cm-roles.yaml. This template creates a CloudFormation stack that includes the service role and instance profile that you need.

newCreateServer

Arguments

:: Text
$sel:engine:CreateServer'

-> Text
$sel:serverName:CreateServer'

-> Text
$sel:instanceProfileArn:CreateServer'

-> Text
$sel:instanceType:CreateServer'

-> Text
$sel:serviceRoleArn:CreateServer'

-> CreateServer

Create a value of CreateServer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:preferredBackupWindow:CreateServer', createServer_preferredBackupWindow - The start time for a one-hour period during which AWS OpsWorks CM backs up application-level data on your server if automated backups are enabled. Valid values must be specified in one of the following formats:

HH:MM for daily backups

DDD:HH:MM for weekly backups

MM must be specified as 00. The specified time is in coordinated universal time (UTC). The default value is a random, daily start time.

Example: 08:00, which represents a daily start time of 08:00 UTC.

Example: Mon:08:00, which represents a start time of every Monday at 08:00 UTC. (8:00 a.m.)

$sel:securityGroupIds:CreateServer', createServer_securityGroupIds - A list of security group IDs to attach to the Amazon EC2 instance. If you add this parameter, the specified security groups must be within the VPC that is specified by SubnetIds.

If you do not specify this parameter, AWS OpsWorks CM creates one new security group that uses TCP ports 22 and 443, open to 0.0.0.0/0 (everyone).

$sel:disableAutomatedBackup:CreateServer', createServer_disableAutomatedBackup - Enable or disable scheduled backups. Valid values are true or false. The default value is true.

$sel:customPrivateKey:CreateServer', createServer_customPrivateKey - A private key in PEM format for connecting to the server by using HTTPS. The private key must not be encrypted; it cannot be protected by a password or passphrase. If you specify a custom private key, you must also specify values for CustomDomain and CustomCertificate.

$sel:engineAttributes:CreateServer', createServer_engineAttributes - Optional engine attributes on a specified server.

Attributes accepted in a Chef createServer request:

CHEF_AUTOMATE_PIVOTAL_KEY: A base64-encoded RSA public key. The corresponding private key is required to access the Chef API. When no CHEF_AUTOMATE_PIVOTAL_KEY is set, a private key is generated and returned in the response.

CHEF_AUTOMATE_ADMIN_PASSWORD: The password for the administrative user in the Chef Automate web-based dashboard. The password length is a minimum of eight characters, and a maximum of 32. The password can contain letters, numbers, and special characters (!/@#$%^&+=_). The password must contain at least one lower case letter, one upper case letter, one number, and one special character. When no CHEF_AUTOMATE_ADMIN_PASSWORD is set, one is generated and returned in the response.

Attributes accepted in a Puppet createServer request:

PUPPET_ADMIN_PASSWORD: To work with the Puppet Enterprise console, a password must use ASCII characters.

PUPPET_R10K_REMOTE: The r10k remote is the URL of your control repository (for example, ssh://git@your.git-repo.com:user/control-repo.git). Specifying an r10k remote opens TCP port 8170.

PUPPET_R10K_PRIVATE_KEY: If you are using a private Git repository, add PUPPET_R10K_PRIVATE_KEY to specify a PEM-encoded private SSH key.

$sel:customDomain:CreateServer', createServer_customDomain - An optional public endpoint of a server, such as https://aws.my-company.com. To access the server, create a CNAME DNS record in your preferred DNS service that points the custom domain to the endpoint that is generated when the server is created (the value of the CreateServer Endpoint attribute). You cannot access the server by using the generated Endpoint value if the server is using a custom domain. If you specify a custom domain, you must also specify values for CustomCertificate and CustomPrivateKey.

$sel:backupId:CreateServer', createServer_backupId - If you specify this field, AWS OpsWorks CM creates the server by using the backup represented by BackupId.

$sel:subnetIds:CreateServer', createServer_subnetIds - The IDs of subnets in which to launch the server EC2 instance.

Amazon EC2-Classic customers: This field is required. All servers must run within a VPC. The VPC must have "Auto Assign Public IP" enabled.

EC2-VPC customers: This field is optional. If you do not specify subnet IDs, your EC2 instances are created in a default subnet that is selected by Amazon EC2. If you specify subnet IDs, the VPC must have "Auto Assign Public IP" enabled.

For more information about supported Amazon EC2 platforms, see Supported Platforms.

$sel:keyPair:CreateServer', createServer_keyPair - The Amazon EC2 key pair to set for the instance. This parameter is optional; if desired, you may specify this parameter to connect to your instances by using SSH.

$sel:associatePublicIpAddress:CreateServer', createServer_associatePublicIpAddress - Associate a public IP address with a server that you are launching. Valid values are true or false. The default value is true.

$sel:engineVersion:CreateServer', createServer_engineVersion - The major release version of the engine that you want to use. For a Chef server, the valid value for EngineVersion is currently 2. For a Puppet server, valid values are 2019 or 2017.

$sel:preferredMaintenanceWindow:CreateServer', createServer_preferredMaintenanceWindow - The start time for a one-hour period each week during which AWS OpsWorks CM performs maintenance on the instance. Valid values must be specified in the following format: DDD:HH:MM. MM must be specified as 00. The specified time is in coordinated universal time (UTC). The default value is a random one-hour period on Tuesday, Wednesday, or Friday. See TimeWindowDefinition for more information.

Example: Mon:08:00, which represents a start time of every Monday at 08:00 UTC. (8:00 a.m.)

$sel:tags:CreateServer', createServer_tags - A map that contains tag keys and tag values to attach to an AWS OpsWorks for Chef Automate or AWS OpsWorks for Puppet Enterprise server.

The key cannot be empty.

The key can be a maximum of 127 characters, and can contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : / @

The value can be a maximum 255 characters, and contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : / @

Leading and trailing white spaces are trimmed from both the key and value.

A maximum of 50 user-applied tags is allowed for any AWS OpsWorks-CM server.

$sel:engineModel:CreateServer', createServer_engineModel - The engine model of the server. Valid values in this release include Monolithic for Puppet and Single for Chef.

$sel:backupRetentionCount:CreateServer', createServer_backupRetentionCount - The number of automated backups that you want to keep. Whenever a new backup is created, AWS OpsWorks CM deletes the oldest backups if this number is exceeded. The default value is 1.

$sel:customCertificate:CreateServer', createServer_customCertificate - A PEM-formatted HTTPS certificate. The value can be be a single, self-signed certificate, or a certificate chain. If you specify a custom certificate, you must also specify values for CustomDomain and CustomPrivateKey. The following are requirements for the CustomCertificate value:

You can provide either a self-signed, custom certificate, or the full certificate chain.

The certificate must be a valid X509 certificate, or a certificate chain in PEM format.

The certificate must be valid at the time of upload. A certificate can't be used before its validity period begins (the certificate's NotBefore date), or after it expires (the certificate's NotAfter date).

The certificate’s common name or subject alternative names (SANs), if present, must match the value of CustomDomain.

The certificate must match the value of CustomPrivateKey.

$sel:engine:CreateServer', createServer_engine - The configuration management engine to use. Valid values include ChefAutomate and Puppet.

$sel:serverName:CreateServer', createServer_serverName - The name of the server. The server name must be unique within your AWS account, within each region. Server names must start with a letter; then letters, numbers, or hyphens (-) are allowed, up to a maximum of 40 characters.

$sel:instanceProfileArn:CreateServer', createServer_instanceProfileArn - The ARN of the instance profile that your Amazon EC2 instances use. Although the AWS OpsWorks console typically creates the instance profile for you, if you are using API commands instead, run the service-role-creation.yaml AWS CloudFormation template, located at https://s3.amazonaws.com/opsworks-cm-us-east-1-prod-default-assets/misc/opsworks-cm-roles.yaml. This template creates a CloudFormation stack that includes the instance profile you need.

$sel:instanceType:CreateServer', createServer_instanceType - The Amazon EC2 instance type to use. For example, m5.large.

$sel:serviceRoleArn:CreateServer', createServer_serviceRoleArn - The service role that the AWS OpsWorks CM service backend uses to work with your account. Although the AWS OpsWorks management console typically creates the service role for you, if you are using the AWS CLI or API commands, run the service-role-creation.yaml AWS CloudFormation template, located at https://s3.amazonaws.com/opsworks-cm-us-east-1-prod-default-assets/misc/opsworks-cm-roles.yaml. This template creates a CloudFormation stack that includes the service role and instance profile that you need.

Request Lenses

createServer_preferredBackupWindow :: Lens' CreateServer (Maybe Text)

The start time for a one-hour period during which AWS OpsWorks CM backs up application-level data on your server if automated backups are enabled. Valid values must be specified in one of the following formats:

HH:MM for daily backups

DDD:HH:MM for weekly backups

MM must be specified as 00. The specified time is in coordinated universal time (UTC). The default value is a random, daily start time.

Example: 08:00, which represents a daily start time of 08:00 UTC.

Example: Mon:08:00, which represents a start time of every Monday at 08:00 UTC. (8:00 a.m.)

createServer_securityGroupIds :: Lens' CreateServer (Maybe [Text])

A list of security group IDs to attach to the Amazon EC2 instance. If you add this parameter, the specified security groups must be within the VPC that is specified by SubnetIds.

If you do not specify this parameter, AWS OpsWorks CM creates one new security group that uses TCP ports 22 and 443, open to 0.0.0.0/0 (everyone).

createServer_disableAutomatedBackup :: Lens' CreateServer (Maybe Bool)

Enable or disable scheduled backups. Valid values are true or false. The default value is true.

createServer_customPrivateKey :: Lens' CreateServer (Maybe Text)

A private key in PEM format for connecting to the server by using HTTPS. The private key must not be encrypted; it cannot be protected by a password or passphrase. If you specify a custom private key, you must also specify values for CustomDomain and CustomCertificate.

createServer_engineAttributes :: Lens' CreateServer (Maybe [EngineAttribute])

Optional engine attributes on a specified server.

Attributes accepted in a Chef createServer request:

CHEF_AUTOMATE_PIVOTAL_KEY: A base64-encoded RSA public key. The corresponding private key is required to access the Chef API. When no CHEF_AUTOMATE_PIVOTAL_KEY is set, a private key is generated and returned in the response.

CHEF_AUTOMATE_ADMIN_PASSWORD: The password for the administrative user in the Chef Automate web-based dashboard. The password length is a minimum of eight characters, and a maximum of 32. The password can contain letters, numbers, and special characters (!/@#$%^&+=_). The password must contain at least one lower case letter, one upper case letter, one number, and one special character. When no CHEF_AUTOMATE_ADMIN_PASSWORD is set, one is generated and returned in the response.

Attributes accepted in a Puppet createServer request:

PUPPET_ADMIN_PASSWORD: To work with the Puppet Enterprise console, a password must use ASCII characters.

PUPPET_R10K_REMOTE: The r10k remote is the URL of your control repository (for example, ssh://git@your.git-repo.com:user/control-repo.git). Specifying an r10k remote opens TCP port 8170.

PUPPET_R10K_PRIVATE_KEY: If you are using a private Git repository, add PUPPET_R10K_PRIVATE_KEY to specify a PEM-encoded private SSH key.

createServer_customDomain :: Lens' CreateServer (Maybe Text)

An optional public endpoint of a server, such as https://aws.my-company.com. To access the server, create a CNAME DNS record in your preferred DNS service that points the custom domain to the endpoint that is generated when the server is created (the value of the CreateServer Endpoint attribute). You cannot access the server by using the generated Endpoint value if the server is using a custom domain. If you specify a custom domain, you must also specify values for CustomCertificate and CustomPrivateKey.

createServer_backupId :: Lens' CreateServer (Maybe Text)

If you specify this field, AWS OpsWorks CM creates the server by using the backup represented by BackupId.

createServer_subnetIds :: Lens' CreateServer (Maybe [Text])

The IDs of subnets in which to launch the server EC2 instance.

Amazon EC2-Classic customers: This field is required. All servers must run within a VPC. The VPC must have "Auto Assign Public IP" enabled.

EC2-VPC customers: This field is optional. If you do not specify subnet IDs, your EC2 instances are created in a default subnet that is selected by Amazon EC2. If you specify subnet IDs, the VPC must have "Auto Assign Public IP" enabled.

For more information about supported Amazon EC2 platforms, see Supported Platforms.

createServer_keyPair :: Lens' CreateServer (Maybe Text)

The Amazon EC2 key pair to set for the instance. This parameter is optional; if desired, you may specify this parameter to connect to your instances by using SSH.

createServer_associatePublicIpAddress :: Lens' CreateServer (Maybe Bool)

Associate a public IP address with a server that you are launching. Valid values are true or false. The default value is true.

createServer_engineVersion :: Lens' CreateServer (Maybe Text)

The major release version of the engine that you want to use. For a Chef server, the valid value for EngineVersion is currently 2. For a Puppet server, valid values are 2019 or 2017.

createServer_preferredMaintenanceWindow :: Lens' CreateServer (Maybe Text)

The start time for a one-hour period each week during which AWS OpsWorks CM performs maintenance on the instance. Valid values must be specified in the following format: DDD:HH:MM. MM must be specified as 00. The specified time is in coordinated universal time (UTC). The default value is a random one-hour period on Tuesday, Wednesday, or Friday. See TimeWindowDefinition for more information.

Example: Mon:08:00, which represents a start time of every Monday at 08:00 UTC. (8:00 a.m.)

createServer_tags :: Lens' CreateServer (Maybe [Tag])

A map that contains tag keys and tag values to attach to an AWS OpsWorks for Chef Automate or AWS OpsWorks for Puppet Enterprise server.

The key cannot be empty.

The key can be a maximum of 127 characters, and can contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : / @

The value can be a maximum 255 characters, and contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : / @

Leading and trailing white spaces are trimmed from both the key and value.

A maximum of 50 user-applied tags is allowed for any AWS OpsWorks-CM server.

createServer_engineModel :: Lens' CreateServer (Maybe Text)

The engine model of the server. Valid values in this release include Monolithic for Puppet and Single for Chef.

createServer_backupRetentionCount :: Lens' CreateServer (Maybe Natural)

The number of automated backups that you want to keep. Whenever a new backup is created, AWS OpsWorks CM deletes the oldest backups if this number is exceeded. The default value is 1.

createServer_customCertificate :: Lens' CreateServer (Maybe Text)

A PEM-formatted HTTPS certificate. The value can be be a single, self-signed certificate, or a certificate chain. If you specify a custom certificate, you must also specify values for CustomDomain and CustomPrivateKey. The following are requirements for the CustomCertificate value:

You can provide either a self-signed, custom certificate, or the full certificate chain.

The certificate must be a valid X509 certificate, or a certificate chain in PEM format.

The certificate must be valid at the time of upload. A certificate can't be used before its validity period begins (the certificate's NotBefore date), or after it expires (the certificate's NotAfter date).

The certificate’s common name or subject alternative names (SANs), if present, must match the value of CustomDomain.

The certificate must match the value of CustomPrivateKey.

createServer_engine :: Lens' CreateServer Text

The configuration management engine to use. Valid values include ChefAutomate and Puppet.

createServer_serverName :: Lens' CreateServer Text

The name of the server. The server name must be unique within your AWS account, within each region. Server names must start with a letter; then letters, numbers, or hyphens (-) are allowed, up to a maximum of 40 characters.

createServer_instanceProfileArn :: Lens' CreateServer Text

The ARN of the instance profile that your Amazon EC2 instances use. Although the AWS OpsWorks console typically creates the instance profile for you, if you are using API commands instead, run the service-role-creation.yaml AWS CloudFormation template, located at https://s3.amazonaws.com/opsworks-cm-us-east-1-prod-default-assets/misc/opsworks-cm-roles.yaml. This template creates a CloudFormation stack that includes the instance profile you need.

createServer_instanceType :: Lens' CreateServer Text

The Amazon EC2 instance type to use. For example, m5.large.

createServer_serviceRoleArn :: Lens' CreateServer Text

The service role that the AWS OpsWorks CM service backend uses to work with your account. Although the AWS OpsWorks management console typically creates the service role for you, if you are using the AWS CLI or API commands, run the service-role-creation.yaml AWS CloudFormation template, located at https://s3.amazonaws.com/opsworks-cm-us-east-1-prod-default-assets/misc/opsworks-cm-roles.yaml. This template creates a CloudFormation stack that includes the service role and instance profile that you need.

Destructuring the Response

data CreateServerResponse

See: newCreateServerResponse smart constructor.

Constructors

CreateServerResponse'

Fields

server :: Maybe Server

The server that is created by the request.

httpStatus :: Int

The response's http status code.

newCreateServerResponse

Arguments

:: Int
$sel:httpStatus:CreateServerResponse'

-> CreateServerResponse

Create a value of CreateServerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:server:CreateServerResponse', createServerResponse_server - The server that is created by the request.

$sel:httpStatus:CreateServerResponse', createServerResponse_httpStatus - The response's http status code.

Response Lenses

createServerResponse_server :: Lens' CreateServerResponse (Maybe Server)

The server that is created by the request.

createServerResponse_httpStatus :: Lens' CreateServerResponse Int

The response's http status code.