Network.AWS.OpsWorksCM.DescribeNodeAssociationStatus

Description

Returns the current status of an existing association or disassociation request.

A ResourceNotFoundException is thrown when no recent association or disassociation request with the specified token is found, or when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Creating a Request

data DescribeNodeAssociationStatus

See: newDescribeNodeAssociationStatus smart constructor.

Constructors

DescribeNodeAssociationStatus'

Fields

nodeAssociationStatusToken :: Text

The token returned in either the AssociateNodeResponse or the DisassociateNodeResponse.

serverName :: Text

The name of the server from which to disassociate the node.

newDescribeNodeAssociationStatus

Arguments

:: Text
$sel:nodeAssociationStatusToken:DescribeNodeAssociationStatus'

-> Text
$sel:serverName:DescribeNodeAssociationStatus'

-> DescribeNodeAssociationStatus

Create a value of DescribeNodeAssociationStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nodeAssociationStatusToken:DescribeNodeAssociationStatus', describeNodeAssociationStatus_nodeAssociationStatusToken - The token returned in either the AssociateNodeResponse or the DisassociateNodeResponse.

$sel:serverName:DescribeNodeAssociationStatus', describeNodeAssociationStatus_serverName - The name of the server from which to disassociate the node.

Request Lenses

describeNodeAssociationStatus_nodeAssociationStatusToken :: Lens' DescribeNodeAssociationStatus Text

The token returned in either the AssociateNodeResponse or the DisassociateNodeResponse.

describeNodeAssociationStatus_serverName :: Lens' DescribeNodeAssociationStatus Text

The name of the server from which to disassociate the node.

Destructuring the Response

data DescribeNodeAssociationStatusResponse

See: newDescribeNodeAssociationStatusResponse smart constructor.

Constructors

DescribeNodeAssociationStatusResponse'

Fields

engineAttributes :: Maybe [EngineAttribute]

Attributes specific to the node association. In Puppet, the attibute PUPPET_NODE_CERT contains the signed certificate (the result of the CSR).

httpStatus :: Int

The response's http status code.

nodeAssociationStatus :: NodeAssociationStatus

The status of the association or disassociation request.

Possible values:

SUCCESS: The association or disassociation succeeded.

FAILED: The association or disassociation failed.

IN_PROGRESS: The association or disassociation is still in progress.

newDescribeNodeAssociationStatusResponse

Arguments

:: Int
$sel:httpStatus:DescribeNodeAssociationStatusResponse'

-> NodeAssociationStatus
$sel:nodeAssociationStatus:DescribeNodeAssociationStatusResponse'

-> DescribeNodeAssociationStatusResponse

Create a value of DescribeNodeAssociationStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineAttributes:DescribeNodeAssociationStatusResponse', describeNodeAssociationStatusResponse_engineAttributes - Attributes specific to the node association. In Puppet, the attibute PUPPET_NODE_CERT contains the signed certificate (the result of the CSR).

$sel:httpStatus:DescribeNodeAssociationStatusResponse', describeNodeAssociationStatusResponse_httpStatus - The response's http status code.

$sel:nodeAssociationStatus:DescribeNodeAssociationStatusResponse', describeNodeAssociationStatusResponse_nodeAssociationStatus - The status of the association or disassociation request.

Possible values:

SUCCESS: The association or disassociation succeeded.

FAILED: The association or disassociation failed.

IN_PROGRESS: The association or disassociation is still in progress.

Response Lenses

describeNodeAssociationStatusResponse_engineAttributes :: Lens' DescribeNodeAssociationStatusResponse (Maybe [EngineAttribute])

Attributes specific to the node association. In Puppet, the attibute PUPPET_NODE_CERT contains the signed certificate (the result of the CSR).

describeNodeAssociationStatusResponse_httpStatus :: Lens' DescribeNodeAssociationStatusResponse Int

The response's http status code.

describeNodeAssociationStatusResponse_nodeAssociationStatus :: Lens' DescribeNodeAssociationStatusResponse NodeAssociationStatus

The status of the association or disassociation request.

Possible values:

SUCCESS: The association or disassociation succeeded.

FAILED: The association or disassociation failed.

IN_PROGRESS: The association or disassociation is still in progress.