Network.AWS.OpsWorksCM.DisassociateNode

Description

Disassociates a node from an AWS OpsWorks CM server, and removes the node from the server's managed nodes. After a node is disassociated, the node key pair is no longer valid for accessing the configuration manager's API. For more information about how to associate a node, see AssociateNode.

A node can can only be disassociated from a server that is in a HEALTHY state. Otherwise, an InvalidStateException is thrown. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Creating a Request

data DisassociateNode

See: newDisassociateNode smart constructor.

Constructors

DisassociateNode'

Fields

engineAttributes :: Maybe [EngineAttribute]

Engine attributes that are used for disassociating the node. No attributes are required for Puppet.

Attributes required in a DisassociateNode request for Chef

CHEF_ORGANIZATION: The Chef organization with which the node was associated. By default only one organization named default can exist.

serverName :: Text

The name of the server from which to disassociate the node.

nodeName :: Text

The name of the client node.

newDisassociateNode

Arguments

:: Text
$sel:serverName:DisassociateNode'

-> Text
$sel:nodeName:DisassociateNode'

-> DisassociateNode

Create a value of DisassociateNode with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineAttributes:DisassociateNode', disassociateNode_engineAttributes - Engine attributes that are used for disassociating the node. No attributes are required for Puppet.

Attributes required in a DisassociateNode request for Chef

CHEF_ORGANIZATION: The Chef organization with which the node was associated. By default only one organization named default can exist.

$sel:serverName:DisassociateNode', disassociateNode_serverName - The name of the server from which to disassociate the node.

$sel:nodeName:DisassociateNode', disassociateNode_nodeName - The name of the client node.

Request Lenses

disassociateNode_engineAttributes :: Lens' DisassociateNode (Maybe [EngineAttribute])

Engine attributes that are used for disassociating the node. No attributes are required for Puppet.

Attributes required in a DisassociateNode request for Chef

CHEF_ORGANIZATION: The Chef organization with which the node was associated. By default only one organization named default can exist.

disassociateNode_serverName :: Lens' DisassociateNode Text

The name of the server from which to disassociate the node.

disassociateNode_nodeName :: Lens' DisassociateNode Text

The name of the client node.

Destructuring the Response

data DisassociateNodeResponse

See: newDisassociateNodeResponse smart constructor.

Constructors

DisassociateNodeResponse'

Fields

nodeAssociationStatusToken :: Maybe Text

Contains a token which can be passed to the DescribeNodeAssociationStatus API call to get the status of the disassociation request.

httpStatus :: Int

The response's http status code.

newDisassociateNodeResponse

Arguments

:: Int
$sel:httpStatus:DisassociateNodeResponse'

-> DisassociateNodeResponse

Create a value of DisassociateNodeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nodeAssociationStatusToken:DisassociateNodeResponse', disassociateNodeResponse_nodeAssociationStatusToken - Contains a token which can be passed to the DescribeNodeAssociationStatus API call to get the status of the disassociation request.

$sel:httpStatus:DisassociateNodeResponse', disassociateNodeResponse_httpStatus - The response's http status code.

Response Lenses

disassociateNodeResponse_nodeAssociationStatusToken :: Lens' DisassociateNodeResponse (Maybe Text)

Contains a token which can be passed to the DescribeNodeAssociationStatus API call to get the status of the disassociation request.

disassociateNodeResponse_httpStatus :: Lens' DisassociateNodeResponse Int

The response's http status code.