Network.AWS.OpsWorksCM.ExportServerEngineAttribute

Description

Exports a specified server engine attribute as a base64-encoded string. For example, you can export user data that you can use in EC2 to associate nodes with a server.

This operation is synchronous.

A ValidationException is raised when parameters of the request are not valid. A ResourceNotFoundException is thrown when the server does not exist. An InvalidStateException is thrown when the server is in any of the following states: CREATING, TERMINATED, FAILED or DELETING.

Creating a Request

data ExportServerEngineAttribute

See: newExportServerEngineAttribute smart constructor.

Constructors

ExportServerEngineAttribute'

Fields

inputAttributes :: Maybe [EngineAttribute]

The list of engine attributes. The list type is EngineAttribute. An EngineAttribute list item is a pair that includes an attribute name and its value. For the Userdata ExportAttributeName, the following are supported engine attribute names.

RunList In Chef, a list of roles or recipes that are run in the specified order. In Puppet, this parameter is ignored.

OrganizationName In Chef, an organization name. AWS OpsWorks for Chef Automate always creates the organization default. In Puppet, this parameter is ignored.

NodeEnvironment In Chef, a node environment (for example, development, staging, or one-box). In Puppet, this parameter is ignored.

NodeClientVersion In Chef, the version of the Chef engine (three numbers separated by dots, such as 13.8.5). If this attribute is empty, OpsWorks for Chef Automate uses the most current version. In Puppet, this parameter is ignored.

exportAttributeName :: Text

The name of the export attribute. Currently, the supported export attribute is Userdata. This exports a user data script that includes parameters and values provided in the InputAttributes list.

serverName :: Text

The name of the server from which you are exporting the attribute.

newExportServerEngineAttribute

Arguments

:: Text
$sel:exportAttributeName:ExportServerEngineAttribute'

-> Text
$sel:serverName:ExportServerEngineAttribute'

-> ExportServerEngineAttribute

Create a value of ExportServerEngineAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inputAttributes:ExportServerEngineAttribute', exportServerEngineAttribute_inputAttributes - The list of engine attributes. The list type is EngineAttribute. An EngineAttribute list item is a pair that includes an attribute name and its value. For the Userdata ExportAttributeName, the following are supported engine attribute names.

RunList In Chef, a list of roles or recipes that are run in the specified order. In Puppet, this parameter is ignored.

OrganizationName In Chef, an organization name. AWS OpsWorks for Chef Automate always creates the organization default. In Puppet, this parameter is ignored.

NodeEnvironment In Chef, a node environment (for example, development, staging, or one-box). In Puppet, this parameter is ignored.

NodeClientVersion In Chef, the version of the Chef engine (three numbers separated by dots, such as 13.8.5). If this attribute is empty, OpsWorks for Chef Automate uses the most current version. In Puppet, this parameter is ignored.

$sel:exportAttributeName:ExportServerEngineAttribute', exportServerEngineAttribute_exportAttributeName - The name of the export attribute. Currently, the supported export attribute is Userdata. This exports a user data script that includes parameters and values provided in the InputAttributes list.

$sel:serverName:ExportServerEngineAttribute', exportServerEngineAttribute_serverName - The name of the server from which you are exporting the attribute.

Request Lenses

exportServerEngineAttribute_inputAttributes :: Lens' ExportServerEngineAttribute (Maybe [EngineAttribute])

The list of engine attributes. The list type is EngineAttribute. An EngineAttribute list item is a pair that includes an attribute name and its value. For the Userdata ExportAttributeName, the following are supported engine attribute names.

RunList In Chef, a list of roles or recipes that are run in the specified order. In Puppet, this parameter is ignored.

OrganizationName In Chef, an organization name. AWS OpsWorks for Chef Automate always creates the organization default. In Puppet, this parameter is ignored.

NodeEnvironment In Chef, a node environment (for example, development, staging, or one-box). In Puppet, this parameter is ignored.

NodeClientVersion In Chef, the version of the Chef engine (three numbers separated by dots, such as 13.8.5). If this attribute is empty, OpsWorks for Chef Automate uses the most current version. In Puppet, this parameter is ignored.

exportServerEngineAttribute_exportAttributeName :: Lens' ExportServerEngineAttribute Text

The name of the export attribute. Currently, the supported export attribute is Userdata. This exports a user data script that includes parameters and values provided in the InputAttributes list.

exportServerEngineAttribute_serverName :: Lens' ExportServerEngineAttribute Text

The name of the server from which you are exporting the attribute.

Destructuring the Response

data ExportServerEngineAttributeResponse

See: newExportServerEngineAttributeResponse smart constructor.

Constructors

ExportServerEngineAttributeResponse'

Fields

engineAttribute :: Maybe EngineAttribute

The requested engine attribute pair with attribute name and value.

serverName :: Maybe Text

The server name used in the request.

httpStatus :: Int

The response's http status code.

newExportServerEngineAttributeResponse

Arguments

:: Int
$sel:httpStatus:ExportServerEngineAttributeResponse'

-> ExportServerEngineAttributeResponse

Create a value of ExportServerEngineAttributeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineAttribute:ExportServerEngineAttributeResponse', exportServerEngineAttributeResponse_engineAttribute - The requested engine attribute pair with attribute name and value.

$sel:serverName:ExportServerEngineAttribute', exportServerEngineAttributeResponse_serverName - The server name used in the request.

$sel:httpStatus:ExportServerEngineAttributeResponse', exportServerEngineAttributeResponse_httpStatus - The response's http status code.

Response Lenses

exportServerEngineAttributeResponse_engineAttribute :: Lens' ExportServerEngineAttributeResponse (Maybe EngineAttribute)

The requested engine attribute pair with attribute name and value.

exportServerEngineAttributeResponse_serverName :: Lens' ExportServerEngineAttributeResponse (Maybe Text)

The server name used in the request.

exportServerEngineAttributeResponse_httpStatus :: Lens' ExportServerEngineAttributeResponse Int

The response's http status code.