Network.AWS.OpsWorksCM.Types

Description

Service Configuration

defaultService :: Service

API version 2016-11-01 of the Amazon OpsWorks CM SDK configuration.

Errors

_InvalidStateException :: AsError a => Getting (First ServiceError) a ServiceError

The resource is in a state that does not allow you to perform a specified action.

_ResourceAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The requested resource cannot be created because it already exists.

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

This occurs when the provided nextToken is not valid.

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

One or more of the provided request parameters are not valid.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The limit of servers or backups has been reached.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The requested resource does not exist, or access was denied.

BackupStatus

newtype BackupStatus

Constructors

BackupStatus'

Fields

fromBackupStatus :: Text

Bundled Patterns

pattern BackupStatus_DELETING :: BackupStatus

pattern BackupStatus_FAILED :: BackupStatus

pattern BackupStatus_IN_PROGRESS :: BackupStatus

pattern BackupStatus_OK :: BackupStatus

BackupType

newtype BackupType

Constructors

BackupType'

Fields

fromBackupType :: Text

Bundled Patterns

pattern BackupType_AUTOMATED :: BackupType

pattern BackupType_MANUAL :: BackupType

MaintenanceStatus

newtype MaintenanceStatus

Constructors

MaintenanceStatus'

Fields

fromMaintenanceStatus :: Text

Bundled Patterns

pattern MaintenanceStatus_FAILED :: MaintenanceStatus

pattern MaintenanceStatus_SUCCESS :: MaintenanceStatus

NodeAssociationStatus

newtype NodeAssociationStatus

The status of the association or disassociation request.

Possible values:

SUCCESS: The association or disassociation succeeded.

FAILED: The association or disassociation failed.

IN_PROGRESS: The association or disassociation is still in progress.

Constructors

NodeAssociationStatus'

Fields

fromNodeAssociationStatus :: Text

Bundled Patterns

pattern NodeAssociationStatus_FAILED :: NodeAssociationStatus

pattern NodeAssociationStatus_IN_PROGRESS :: NodeAssociationStatus

pattern NodeAssociationStatus_SUCCESS :: NodeAssociationStatus

ServerStatus

newtype ServerStatus

Constructors

ServerStatus'

Fields

fromServerStatus :: Text

Bundled Patterns

pattern ServerStatus_BACKING_UP :: ServerStatus

pattern ServerStatus_CONNECTION_LOST :: ServerStatus

pattern ServerStatus_CREATING :: ServerStatus

pattern ServerStatus_DELETING :: ServerStatus

pattern ServerStatus_FAILED :: ServerStatus

pattern ServerStatus_HEALTHY :: ServerStatus

pattern ServerStatus_MODIFYING :: ServerStatus

pattern ServerStatus_RESTORING :: ServerStatus

pattern ServerStatus_RUNNING :: ServerStatus

pattern ServerStatus_SETUP :: ServerStatus

pattern ServerStatus_TERMINATED :: ServerStatus

pattern ServerStatus_UNDER_MAINTENANCE :: ServerStatus

pattern ServerStatus_UNHEALTHY :: ServerStatus

AccountAttribute

data AccountAttribute

Stores account attributes.

See: newAccountAttribute smart constructor.

Constructors

AccountAttribute'

Fields

used :: Maybe Int

The current usage, such as the current number of servers that are associated with the account.

name :: Maybe Text

The attribute name. The following are supported attribute names.

ServerLimit: The number of current servers/maximum number of servers allowed. By default, you can have a maximum of 10 servers.

ManualBackupLimit: The number of current manual backups/maximum number of backups allowed. By default, you can have a maximum of 50 manual backups saved.

maximum :: Maybe Int

The maximum allowed value.

newAccountAttribute :: AccountAttribute

Create a value of AccountAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:used:AccountAttribute', accountAttribute_used - The current usage, such as the current number of servers that are associated with the account.

$sel:name:AccountAttribute', accountAttribute_name - The attribute name. The following are supported attribute names.

ServerLimit: The number of current servers/maximum number of servers allowed. By default, you can have a maximum of 10 servers.

ManualBackupLimit: The number of current manual backups/maximum number of backups allowed. By default, you can have a maximum of 50 manual backups saved.

$sel:maximum:AccountAttribute', accountAttribute_maximum - The maximum allowed value.

accountAttribute_used :: Lens' AccountAttribute (Maybe Int)

The current usage, such as the current number of servers that are associated with the account.

accountAttribute_name :: Lens' AccountAttribute (Maybe Text)

The attribute name. The following are supported attribute names.

ServerLimit: The number of current servers/maximum number of servers allowed. By default, you can have a maximum of 10 servers.

ManualBackupLimit: The number of current manual backups/maximum number of backups allowed. By default, you can have a maximum of 50 manual backups saved.

accountAttribute_maximum :: Lens' AccountAttribute (Maybe Int)

The maximum allowed value.

Backup

data Backup

Describes a single backup.

See: newBackup smart constructor.

Constructors

Backup'

Fields

instanceProfileArn :: Maybe Text

The EC2 instance profile ARN that is obtained from the server when the backup is created. Because this value is stored, you are not required to provide the InstanceProfileArn again if you restore a backup.

preferredBackupWindow :: Maybe Text

The preferred backup period that is obtained from the server when the backup is created.

securityGroupIds :: Maybe [Text]

The security group IDs that are obtained from the server when the backup is created.

serviceRoleArn :: Maybe Text

The service role ARN that is obtained from the server when the backup is created.

status :: Maybe BackupStatus

The status of a backup while in progress.

userArn :: Maybe Text

The IAM user ARN of the requester for manual backups. This field is empty for automated backups.

instanceType :: Maybe Text

The instance type that is obtained from the server when the backup is created.

backupType :: Maybe BackupType

The backup type. Valid values are automated or manual.

s3DataUrl :: Maybe Text

This field is deprecated and is no longer used.

backupArn :: Maybe Text

The ARN of the backup.

backupId :: Maybe Text

The generated ID of the backup. Example: myServerName-yyyyMMddHHmmssSSS

createdAt :: Maybe POSIX

The time stamp when the backup was created in the database. Example: 2016-07-29T13:38:47.520Z

subnetIds :: Maybe [Text]

The subnet IDs that are obtained from the server when the backup is created.

s3DataSize :: Maybe Int

This field is deprecated and is no longer used.

keyPair :: Maybe Text

The key pair that is obtained from the server when the backup is created.

serverName :: Maybe Text

The name of the server from which the backup was made.

s3LogUrl :: Maybe Text

The Amazon S3 URL of the backup's log file.

engineVersion :: Maybe Text

The engine version that is obtained from the server when the backup is created.

preferredMaintenanceWindow :: Maybe Text

The preferred maintenance period that is obtained from the server when the backup is created.

toolsVersion :: Maybe Text

The version of AWS OpsWorks CM-specific tools that is obtained from the server when the backup is created.

engineModel :: Maybe Text

The engine model that is obtained from the server when the backup is created.

engine :: Maybe Text

The engine type that is obtained from the server when the backup is created.

description :: Maybe Text

A user-provided description for a manual backup. This field is empty for automated backups.

statusDescription :: Maybe Text

An informational message about backup status.

newBackup :: Backup

Create a value of Backup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceProfileArn:Backup', backup_instanceProfileArn - The EC2 instance profile ARN that is obtained from the server when the backup is created. Because this value is stored, you are not required to provide the InstanceProfileArn again if you restore a backup.

$sel:preferredBackupWindow:Backup', backup_preferredBackupWindow - The preferred backup period that is obtained from the server when the backup is created.

$sel:securityGroupIds:Backup', backup_securityGroupIds - The security group IDs that are obtained from the server when the backup is created.

$sel:serviceRoleArn:Backup', backup_serviceRoleArn - The service role ARN that is obtained from the server when the backup is created.

$sel:status:Backup', backup_status - The status of a backup while in progress.

$sel:userArn:Backup', backup_userArn - The IAM user ARN of the requester for manual backups. This field is empty for automated backups.

$sel:instanceType:Backup', backup_instanceType - The instance type that is obtained from the server when the backup is created.

$sel:backupType:Backup', backup_backupType - The backup type. Valid values are automated or manual.

$sel:s3DataUrl:Backup', backup_s3DataUrl - This field is deprecated and is no longer used.

$sel:backupArn:Backup', backup_backupArn - The ARN of the backup.

$sel:backupId:Backup', backup_backupId - The generated ID of the backup. Example: myServerName-yyyyMMddHHmmssSSS

$sel:createdAt:Backup', backup_createdAt - The time stamp when the backup was created in the database. Example: 2016-07-29T13:38:47.520Z

$sel:subnetIds:Backup', backup_subnetIds - The subnet IDs that are obtained from the server when the backup is created.

$sel:s3DataSize:Backup', backup_s3DataSize - This field is deprecated and is no longer used.

$sel:keyPair:Backup', backup_keyPair - The key pair that is obtained from the server when the backup is created.

$sel:serverName:Backup', backup_serverName - The name of the server from which the backup was made.

$sel:s3LogUrl:Backup', backup_s3LogUrl - The Amazon S3 URL of the backup's log file.

$sel:engineVersion:Backup', backup_engineVersion - The engine version that is obtained from the server when the backup is created.

$sel:preferredMaintenanceWindow:Backup', backup_preferredMaintenanceWindow - The preferred maintenance period that is obtained from the server when the backup is created.

$sel:toolsVersion:Backup', backup_toolsVersion - The version of AWS OpsWorks CM-specific tools that is obtained from the server when the backup is created.

$sel:engineModel:Backup', backup_engineModel - The engine model that is obtained from the server when the backup is created.

$sel:engine:Backup', backup_engine - The engine type that is obtained from the server when the backup is created.

$sel:description:Backup', backup_description - A user-provided description for a manual backup. This field is empty for automated backups.

$sel:statusDescription:Backup', backup_statusDescription - An informational message about backup status.

backup_instanceProfileArn :: Lens' Backup (Maybe Text)

The EC2 instance profile ARN that is obtained from the server when the backup is created. Because this value is stored, you are not required to provide the InstanceProfileArn again if you restore a backup.

backup_preferredBackupWindow :: Lens' Backup (Maybe Text)

The preferred backup period that is obtained from the server when the backup is created.

backup_securityGroupIds :: Lens' Backup (Maybe [Text])

The security group IDs that are obtained from the server when the backup is created.

backup_serviceRoleArn :: Lens' Backup (Maybe Text)

The service role ARN that is obtained from the server when the backup is created.

backup_status :: Lens' Backup (Maybe BackupStatus)

The status of a backup while in progress.

backup_userArn :: Lens' Backup (Maybe Text)

The IAM user ARN of the requester for manual backups. This field is empty for automated backups.

backup_instanceType :: Lens' Backup (Maybe Text)

The instance type that is obtained from the server when the backup is created.

backup_backupType :: Lens' Backup (Maybe BackupType)

The backup type. Valid values are automated or manual.

backup_s3DataUrl :: Lens' Backup (Maybe Text)

This field is deprecated and is no longer used.

backup_backupArn :: Lens' Backup (Maybe Text)

The ARN of the backup.

backup_backupId :: Lens' Backup (Maybe Text)

The generated ID of the backup. Example: myServerName-yyyyMMddHHmmssSSS

backup_createdAt :: Lens' Backup (Maybe UTCTime)

The time stamp when the backup was created in the database. Example: 2016-07-29T13:38:47.520Z

backup_subnetIds :: Lens' Backup (Maybe [Text])

The subnet IDs that are obtained from the server when the backup is created.

backup_s3DataSize :: Lens' Backup (Maybe Int)

This field is deprecated and is no longer used.

backup_keyPair :: Lens' Backup (Maybe Text)

The key pair that is obtained from the server when the backup is created.

backup_serverName :: Lens' Backup (Maybe Text)

The name of the server from which the backup was made.

backup_s3LogUrl :: Lens' Backup (Maybe Text)

The Amazon S3 URL of the backup's log file.

backup_engineVersion :: Lens' Backup (Maybe Text)

The engine version that is obtained from the server when the backup is created.

backup_preferredMaintenanceWindow :: Lens' Backup (Maybe Text)

The preferred maintenance period that is obtained from the server when the backup is created.

backup_toolsVersion :: Lens' Backup (Maybe Text)

The version of AWS OpsWorks CM-specific tools that is obtained from the server when the backup is created.

backup_engineModel :: Lens' Backup (Maybe Text)

The engine model that is obtained from the server when the backup is created.

backup_engine :: Lens' Backup (Maybe Text)

The engine type that is obtained from the server when the backup is created.

backup_description :: Lens' Backup (Maybe Text)

A user-provided description for a manual backup. This field is empty for automated backups.

backup_statusDescription :: Lens' Backup (Maybe Text)

An informational message about backup status.

EngineAttribute

data EngineAttribute

A name and value pair that is specific to the engine of the server.

See: newEngineAttribute smart constructor.

Constructors

EngineAttribute'

Fields

name :: Maybe Text

The name of the engine attribute.

value :: Maybe (Sensitive Text)

The value of the engine attribute.

newEngineAttribute :: EngineAttribute

Create a value of EngineAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:EngineAttribute', engineAttribute_name - The name of the engine attribute.

$sel:value:EngineAttribute', engineAttribute_value - The value of the engine attribute.

engineAttribute_name :: Lens' EngineAttribute (Maybe Text)

The name of the engine attribute.

engineAttribute_value :: Lens' EngineAttribute (Maybe Text)

The value of the engine attribute.

Server

data Server

Describes a configuration management server.

See: newServer smart constructor.

Constructors

Server'

Fields

instanceProfileArn :: Maybe Text

The instance profile ARN of the server.

preferredBackupWindow :: Maybe Text

The preferred backup period specified for the server.

securityGroupIds :: Maybe [Text]

The security group IDs for the server, as specified in the CloudFormation stack. These might not be the same security groups that are shown in the EC2 console.

serviceRoleArn :: Maybe Text

The service role ARN used to create the server.

status :: Maybe ServerStatus

The server's status. This field displays the states of actions in progress, such as creating, running, or backing up the server, as well as the server's health state.

disableAutomatedBackup :: Maybe Bool

Disables automated backups. The number of stored backups is dependent on the value of PreferredBackupCount.

engineAttributes :: Maybe [EngineAttribute]

The response of a createServer() request returns the master credential to access the server in EngineAttributes. These credentials are not stored by AWS OpsWorks CM; they are returned only as part of the result of createServer().

Attributes returned in a createServer response for Chef

CHEF_AUTOMATE_PIVOTAL_KEY: A base64-encoded RSA private key that is generated by AWS OpsWorks for Chef Automate. This private key is required to access the Chef API.

CHEF_STARTER_KIT: A base64-encoded ZIP file. The ZIP file contains a Chef starter kit, which includes a README, a configuration file, and the required RSA private key. Save this file, unzip it, and then change to the directory where you've unzipped the file contents. From this directory, you can run Knife commands.

Attributes returned in a createServer response for Puppet

PUPPET_STARTER_KIT: A base64-encoded ZIP file. The ZIP file contains a Puppet starter kit, including a README and a required private key. Save this file, unzip it, and then change to the directory where you've unzipped the file contents.

PUPPET_ADMIN_PASSWORD: An administrator password that you can use to sign in to the Puppet Enterprise console after the server is online.

instanceType :: Maybe Text

The instance type for the server, as specified in the CloudFormation stack. This might not be the same instance type that is shown in the EC2 console.

customDomain :: Maybe Text

An optional public endpoint of a server, such as https://aws.my-company.com. You cannot access the server by using the Endpoint value if the server has a CustomDomain specified.

createdAt :: Maybe POSIX

Time stamp of server creation. Example 2016-07-29T13:38:47.520Z

subnetIds :: Maybe [Text]

The subnet IDs specified in a CreateServer request.

keyPair :: Maybe Text

The key pair associated with the server.

associatePublicIpAddress :: Maybe Bool

Associate a public IP address with a server that you are launching.

serverName :: Maybe Text

The name of the server.

engineVersion :: Maybe Text

The engine version of the server. For a Chef server, the valid value for EngineVersion is currently 2. For a Puppet server, specify either 2019 or 2017.

preferredMaintenanceWindow :: Maybe Text

The preferred maintenance period specified for the server.

cloudFormationStackArn :: Maybe Text

The ARN of the CloudFormation stack that was used to create the server.

maintenanceStatus :: Maybe MaintenanceStatus

The status of the most recent server maintenance run. Shows SUCCESS or FAILED.

engineModel :: Maybe Text

The engine model of the server. Valid values in this release include Monolithic for Puppet and Single for Chef.

backupRetentionCount :: Maybe Int

The number of automated backups to keep.

engine :: Maybe Text

The engine type of the server. Valid values in this release include ChefAutomate and Puppet.

endpoint :: Maybe Text

A DNS name that can be used to access the engine. Example: myserver-asdfghjkl.us-east-1.opsworks.io. You cannot access the server by using the Endpoint value if the server has a CustomDomain specified.

statusReason :: Maybe Text

Depending on the server status, this field has either a human-readable message (such as a create or backup error), or an escaped block of JSON (used for health check results).

serverArn :: Maybe Text

The ARN of the server.

newServer :: Server

Create a value of Server with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceProfileArn:Server', server_instanceProfileArn - The instance profile ARN of the server.

$sel:preferredBackupWindow:Server', server_preferredBackupWindow - The preferred backup period specified for the server.

$sel:securityGroupIds:Server', server_securityGroupIds - The security group IDs for the server, as specified in the CloudFormation stack. These might not be the same security groups that are shown in the EC2 console.

$sel:serviceRoleArn:Server', server_serviceRoleArn - The service role ARN used to create the server.

$sel:status:Server', server_status - The server's status. This field displays the states of actions in progress, such as creating, running, or backing up the server, as well as the server's health state.

$sel:disableAutomatedBackup:Server', server_disableAutomatedBackup - Disables automated backups. The number of stored backups is dependent on the value of PreferredBackupCount.

$sel:engineAttributes:Server', server_engineAttributes - The response of a createServer() request returns the master credential to access the server in EngineAttributes. These credentials are not stored by AWS OpsWorks CM; they are returned only as part of the result of createServer().

Attributes returned in a createServer response for Chef

CHEF_AUTOMATE_PIVOTAL_KEY: A base64-encoded RSA private key that is generated by AWS OpsWorks for Chef Automate. This private key is required to access the Chef API.

CHEF_STARTER_KIT: A base64-encoded ZIP file. The ZIP file contains a Chef starter kit, which includes a README, a configuration file, and the required RSA private key. Save this file, unzip it, and then change to the directory where you've unzipped the file contents. From this directory, you can run Knife commands.

Attributes returned in a createServer response for Puppet

PUPPET_STARTER_KIT: A base64-encoded ZIP file. The ZIP file contains a Puppet starter kit, including a README and a required private key. Save this file, unzip it, and then change to the directory where you've unzipped the file contents.

PUPPET_ADMIN_PASSWORD: An administrator password that you can use to sign in to the Puppet Enterprise console after the server is online.

$sel:instanceType:Server', server_instanceType - The instance type for the server, as specified in the CloudFormation stack. This might not be the same instance type that is shown in the EC2 console.

$sel:customDomain:Server', server_customDomain - An optional public endpoint of a server, such as https://aws.my-company.com. You cannot access the server by using the Endpoint value if the server has a CustomDomain specified.

$sel:createdAt:Server', server_createdAt - Time stamp of server creation. Example 2016-07-29T13:38:47.520Z

$sel:subnetIds:Server', server_subnetIds - The subnet IDs specified in a CreateServer request.

$sel:keyPair:Server', server_keyPair - The key pair associated with the server.

$sel:associatePublicIpAddress:Server', server_associatePublicIpAddress - Associate a public IP address with a server that you are launching.

$sel:serverName:Server', server_serverName - The name of the server.

$sel:engineVersion:Server', server_engineVersion - The engine version of the server. For a Chef server, the valid value for EngineVersion is currently 2. For a Puppet server, specify either 2019 or 2017.

$sel:preferredMaintenanceWindow:Server', server_preferredMaintenanceWindow - The preferred maintenance period specified for the server.

$sel:cloudFormationStackArn:Server', server_cloudFormationStackArn - The ARN of the CloudFormation stack that was used to create the server.

$sel:maintenanceStatus:Server', server_maintenanceStatus - The status of the most recent server maintenance run. Shows SUCCESS or FAILED.

$sel:engineModel:Server', server_engineModel - The engine model of the server. Valid values in this release include Monolithic for Puppet and Single for Chef.

$sel:backupRetentionCount:Server', server_backupRetentionCount - The number of automated backups to keep.

$sel:engine:Server', server_engine - The engine type of the server. Valid values in this release include ChefAutomate and Puppet.

$sel:endpoint:Server', server_endpoint - A DNS name that can be used to access the engine. Example: myserver-asdfghjkl.us-east-1.opsworks.io. You cannot access the server by using the Endpoint value if the server has a CustomDomain specified.

$sel:statusReason:Server', server_statusReason - Depending on the server status, this field has either a human-readable message (such as a create or backup error), or an escaped block of JSON (used for health check results).

$sel:serverArn:Server', server_serverArn - The ARN of the server.

server_instanceProfileArn :: Lens' Server (Maybe Text)

The instance profile ARN of the server.

server_preferredBackupWindow :: Lens' Server (Maybe Text)

The preferred backup period specified for the server.

server_securityGroupIds :: Lens' Server (Maybe [Text])

The security group IDs for the server, as specified in the CloudFormation stack. These might not be the same security groups that are shown in the EC2 console.

server_serviceRoleArn :: Lens' Server (Maybe Text)

The service role ARN used to create the server.

server_status :: Lens' Server (Maybe ServerStatus)

The server's status. This field displays the states of actions in progress, such as creating, running, or backing up the server, as well as the server's health state.

server_disableAutomatedBackup :: Lens' Server (Maybe Bool)

Disables automated backups. The number of stored backups is dependent on the value of PreferredBackupCount.

server_engineAttributes :: Lens' Server (Maybe [EngineAttribute])

The response of a createServer() request returns the master credential to access the server in EngineAttributes. These credentials are not stored by AWS OpsWorks CM; they are returned only as part of the result of createServer().

Attributes returned in a createServer response for Chef

CHEF_AUTOMATE_PIVOTAL_KEY: A base64-encoded RSA private key that is generated by AWS OpsWorks for Chef Automate. This private key is required to access the Chef API.

CHEF_STARTER_KIT: A base64-encoded ZIP file. The ZIP file contains a Chef starter kit, which includes a README, a configuration file, and the required RSA private key. Save this file, unzip it, and then change to the directory where you've unzipped the file contents. From this directory, you can run Knife commands.

Attributes returned in a createServer response for Puppet

PUPPET_STARTER_KIT: A base64-encoded ZIP file. The ZIP file contains a Puppet starter kit, including a README and a required private key. Save this file, unzip it, and then change to the directory where you've unzipped the file contents.

PUPPET_ADMIN_PASSWORD: An administrator password that you can use to sign in to the Puppet Enterprise console after the server is online.

server_instanceType :: Lens' Server (Maybe Text)

The instance type for the server, as specified in the CloudFormation stack. This might not be the same instance type that is shown in the EC2 console.

server_customDomain :: Lens' Server (Maybe Text)

An optional public endpoint of a server, such as https://aws.my-company.com. You cannot access the server by using the Endpoint value if the server has a CustomDomain specified.

server_createdAt :: Lens' Server (Maybe UTCTime)

Time stamp of server creation. Example 2016-07-29T13:38:47.520Z

server_subnetIds :: Lens' Server (Maybe [Text])

The subnet IDs specified in a CreateServer request.

server_keyPair :: Lens' Server (Maybe Text)

The key pair associated with the server.

server_associatePublicIpAddress :: Lens' Server (Maybe Bool)

Associate a public IP address with a server that you are launching.

server_serverName :: Lens' Server (Maybe Text)

The name of the server.

server_engineVersion :: Lens' Server (Maybe Text)

The engine version of the server. For a Chef server, the valid value for EngineVersion is currently 2. For a Puppet server, specify either 2019 or 2017.

server_preferredMaintenanceWindow :: Lens' Server (Maybe Text)

The preferred maintenance period specified for the server.

server_cloudFormationStackArn :: Lens' Server (Maybe Text)

The ARN of the CloudFormation stack that was used to create the server.

server_maintenanceStatus :: Lens' Server (Maybe MaintenanceStatus)

The status of the most recent server maintenance run. Shows SUCCESS or FAILED.

server_engineModel :: Lens' Server (Maybe Text)

The engine model of the server. Valid values in this release include Monolithic for Puppet and Single for Chef.

server_backupRetentionCount :: Lens' Server (Maybe Int)

The number of automated backups to keep.

server_engine :: Lens' Server (Maybe Text)

The engine type of the server. Valid values in this release include ChefAutomate and Puppet.

server_endpoint :: Lens' Server (Maybe Text)

A DNS name that can be used to access the engine. Example: myserver-asdfghjkl.us-east-1.opsworks.io. You cannot access the server by using the Endpoint value if the server has a CustomDomain specified.

server_statusReason :: Lens' Server (Maybe Text)

Depending on the server status, this field has either a human-readable message (such as a create or backup error), or an escaped block of JSON (used for health check results).

server_serverArn :: Lens' Server (Maybe Text)

The ARN of the server.

ServerEvent

data ServerEvent

An event that is related to the server, such as the start of maintenance or backup.

See: newServerEvent smart constructor.

Constructors

ServerEvent'

Fields

logUrl :: Maybe Text

The Amazon S3 URL of the event's log file.

message :: Maybe Text

A human-readable informational or status message.

createdAt :: Maybe POSIX

The time when the event occurred.

serverName :: Maybe Text

The name of the server on or for which the event occurred.

newServerEvent :: ServerEvent

Create a value of ServerEvent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:logUrl:ServerEvent', serverEvent_logUrl - The Amazon S3 URL of the event's log file.

$sel:message:ServerEvent', serverEvent_message - A human-readable informational or status message.

$sel:createdAt:ServerEvent', serverEvent_createdAt - The time when the event occurred.

$sel:serverName:ServerEvent', serverEvent_serverName - The name of the server on or for which the event occurred.

serverEvent_logUrl :: Lens' ServerEvent (Maybe Text)

The Amazon S3 URL of the event's log file.

serverEvent_message :: Lens' ServerEvent (Maybe Text)

A human-readable informational or status message.

serverEvent_createdAt :: Lens' ServerEvent (Maybe UTCTime)

The time when the event occurred.

serverEvent_serverName :: Lens' ServerEvent (Maybe Text)

The name of the server on or for which the event occurred.

Tag

data Tag

A map that contains tag keys and tag values to attach to an AWS OpsWorks for Chef Automate or AWS OpsWorks for Puppet Enterprise server. Leading and trailing white spaces are trimmed from both the key and value. A maximum of 50 user-applied tags is allowed for tag-supported AWS OpsWorks-CM resources.

See: newTag smart constructor.

Constructors

Tag'

Fields

key :: Text

A tag key, such as Stage or Name. A tag key cannot be empty. The key can be a maximum of 127 characters, and can contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : /

value :: Text

An optional tag value, such as Production or test-owcm-server. The value can be a maximum of 255 characters, and contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : /

newTag

Arguments

:: Text
$sel:key:Tag'

-> Text
$sel:value:Tag'

-> Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:key:Tag', tag_key - A tag key, such as Stage or Name. A tag key cannot be empty. The key can be a maximum of 127 characters, and can contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : /

$sel:value:Tag', tag_value - An optional tag value, such as Production or test-owcm-server. The value can be a maximum of 255 characters, and contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : /

tag_key :: Lens' Tag Text

A tag key, such as Stage or Name. A tag key cannot be empty. The key can be a maximum of 127 characters, and can contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : /

tag_value :: Lens' Tag Text

An optional tag value, such as Production or test-owcm-server. The value can be a maximum of 255 characters, and contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : /