Network.AWS.OpsWorksCM.Types.Backup

Description

Documentation

data Backup

Describes a single backup.

See: newBackup smart constructor.

Constructors

Backup'

Fields

instanceProfileArn :: Maybe Text

The EC2 instance profile ARN that is obtained from the server when the backup is created. Because this value is stored, you are not required to provide the InstanceProfileArn again if you restore a backup.

preferredBackupWindow :: Maybe Text

The preferred backup period that is obtained from the server when the backup is created.

securityGroupIds :: Maybe [Text]

The security group IDs that are obtained from the server when the backup is created.

serviceRoleArn :: Maybe Text

The service role ARN that is obtained from the server when the backup is created.

status :: Maybe BackupStatus

The status of a backup while in progress.

userArn :: Maybe Text

The IAM user ARN of the requester for manual backups. This field is empty for automated backups.

instanceType :: Maybe Text

The instance type that is obtained from the server when the backup is created.

backupType :: Maybe BackupType

The backup type. Valid values are automated or manual.

s3DataUrl :: Maybe Text

This field is deprecated and is no longer used.

backupArn :: Maybe Text

The ARN of the backup.

backupId :: Maybe Text

The generated ID of the backup. Example: myServerName-yyyyMMddHHmmssSSS

createdAt :: Maybe POSIX

The time stamp when the backup was created in the database. Example: 2016-07-29T13:38:47.520Z

subnetIds :: Maybe [Text]

The subnet IDs that are obtained from the server when the backup is created.

s3DataSize :: Maybe Int

This field is deprecated and is no longer used.

keyPair :: Maybe Text

The key pair that is obtained from the server when the backup is created.

serverName :: Maybe Text

The name of the server from which the backup was made.

s3LogUrl :: Maybe Text

The Amazon S3 URL of the backup's log file.

engineVersion :: Maybe Text

The engine version that is obtained from the server when the backup is created.

preferredMaintenanceWindow :: Maybe Text

The preferred maintenance period that is obtained from the server when the backup is created.

toolsVersion :: Maybe Text

The version of AWS OpsWorks CM-specific tools that is obtained from the server when the backup is created.

engineModel :: Maybe Text

The engine model that is obtained from the server when the backup is created.

engine :: Maybe Text

The engine type that is obtained from the server when the backup is created.

description :: Maybe Text

A user-provided description for a manual backup. This field is empty for automated backups.

statusDescription :: Maybe Text

An informational message about backup status.

newBackup :: Backup

Create a value of Backup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceProfileArn:Backup', backup_instanceProfileArn - The EC2 instance profile ARN that is obtained from the server when the backup is created. Because this value is stored, you are not required to provide the InstanceProfileArn again if you restore a backup.

$sel:preferredBackupWindow:Backup', backup_preferredBackupWindow - The preferred backup period that is obtained from the server when the backup is created.

$sel:securityGroupIds:Backup', backup_securityGroupIds - The security group IDs that are obtained from the server when the backup is created.

$sel:serviceRoleArn:Backup', backup_serviceRoleArn - The service role ARN that is obtained from the server when the backup is created.

$sel:status:Backup', backup_status - The status of a backup while in progress.

$sel:userArn:Backup', backup_userArn - The IAM user ARN of the requester for manual backups. This field is empty for automated backups.

$sel:instanceType:Backup', backup_instanceType - The instance type that is obtained from the server when the backup is created.

$sel:backupType:Backup', backup_backupType - The backup type. Valid values are automated or manual.

$sel:s3DataUrl:Backup', backup_s3DataUrl - This field is deprecated and is no longer used.

$sel:backupArn:Backup', backup_backupArn - The ARN of the backup.

$sel:backupId:Backup', backup_backupId - The generated ID of the backup. Example: myServerName-yyyyMMddHHmmssSSS

$sel:createdAt:Backup', backup_createdAt - The time stamp when the backup was created in the database. Example: 2016-07-29T13:38:47.520Z

$sel:subnetIds:Backup', backup_subnetIds - The subnet IDs that are obtained from the server when the backup is created.

$sel:s3DataSize:Backup', backup_s3DataSize - This field is deprecated and is no longer used.

$sel:keyPair:Backup', backup_keyPair - The key pair that is obtained from the server when the backup is created.

$sel:serverName:Backup', backup_serverName - The name of the server from which the backup was made.

$sel:s3LogUrl:Backup', backup_s3LogUrl - The Amazon S3 URL of the backup's log file.

$sel:engineVersion:Backup', backup_engineVersion - The engine version that is obtained from the server when the backup is created.

$sel:preferredMaintenanceWindow:Backup', backup_preferredMaintenanceWindow - The preferred maintenance period that is obtained from the server when the backup is created.

$sel:toolsVersion:Backup', backup_toolsVersion - The version of AWS OpsWorks CM-specific tools that is obtained from the server when the backup is created.

$sel:engineModel:Backup', backup_engineModel - The engine model that is obtained from the server when the backup is created.

$sel:engine:Backup', backup_engine - The engine type that is obtained from the server when the backup is created.

$sel:description:Backup', backup_description - A user-provided description for a manual backup. This field is empty for automated backups.

$sel:statusDescription:Backup', backup_statusDescription - An informational message about backup status.

backup_instanceProfileArn :: Lens' Backup (Maybe Text)

The EC2 instance profile ARN that is obtained from the server when the backup is created. Because this value is stored, you are not required to provide the InstanceProfileArn again if you restore a backup.

backup_preferredBackupWindow :: Lens' Backup (Maybe Text)

The preferred backup period that is obtained from the server when the backup is created.

backup_securityGroupIds :: Lens' Backup (Maybe [Text])

The security group IDs that are obtained from the server when the backup is created.

backup_serviceRoleArn :: Lens' Backup (Maybe Text)

The service role ARN that is obtained from the server when the backup is created.

backup_status :: Lens' Backup (Maybe BackupStatus)

The status of a backup while in progress.

backup_userArn :: Lens' Backup (Maybe Text)

The IAM user ARN of the requester for manual backups. This field is empty for automated backups.

backup_instanceType :: Lens' Backup (Maybe Text)

The instance type that is obtained from the server when the backup is created.

backup_backupType :: Lens' Backup (Maybe BackupType)

The backup type. Valid values are automated or manual.

backup_s3DataUrl :: Lens' Backup (Maybe Text)

This field is deprecated and is no longer used.

backup_backupArn :: Lens' Backup (Maybe Text)

The ARN of the backup.

backup_backupId :: Lens' Backup (Maybe Text)

The generated ID of the backup. Example: myServerName-yyyyMMddHHmmssSSS

backup_createdAt :: Lens' Backup (Maybe UTCTime)

The time stamp when the backup was created in the database. Example: 2016-07-29T13:38:47.520Z

backup_subnetIds :: Lens' Backup (Maybe [Text])

The subnet IDs that are obtained from the server when the backup is created.

backup_s3DataSize :: Lens' Backup (Maybe Int)

This field is deprecated and is no longer used.

backup_keyPair :: Lens' Backup (Maybe Text)

The key pair that is obtained from the server when the backup is created.

backup_serverName :: Lens' Backup (Maybe Text)

The name of the server from which the backup was made.

backup_s3LogUrl :: Lens' Backup (Maybe Text)

The Amazon S3 URL of the backup's log file.

backup_engineVersion :: Lens' Backup (Maybe Text)

The engine version that is obtained from the server when the backup is created.

backup_preferredMaintenanceWindow :: Lens' Backup (Maybe Text)

The preferred maintenance period that is obtained from the server when the backup is created.

backup_toolsVersion :: Lens' Backup (Maybe Text)

The version of AWS OpsWorks CM-specific tools that is obtained from the server when the backup is created.

backup_engineModel :: Lens' Backup (Maybe Text)

The engine model that is obtained from the server when the backup is created.

backup_engine :: Lens' Backup (Maybe Text)

The engine type that is obtained from the server when the backup is created.

backup_description :: Lens' Backup (Maybe Text)

A user-provided description for a manual backup. This field is empty for automated backups.

backup_statusDescription :: Lens' Backup (Maybe Text)

An informational message about backup status.