Network.AWS.OpsWorksCM.Types.Server

Description

Documentation

data Server

Describes a configuration management server.

See: newServer smart constructor.

Constructors

Server'

Fields

instanceProfileArn :: Maybe Text

The instance profile ARN of the server.

preferredBackupWindow :: Maybe Text

The preferred backup period specified for the server.

securityGroupIds :: Maybe [Text]

The security group IDs for the server, as specified in the CloudFormation stack. These might not be the same security groups that are shown in the EC2 console.

serviceRoleArn :: Maybe Text

The service role ARN used to create the server.

status :: Maybe ServerStatus

The server's status. This field displays the states of actions in progress, such as creating, running, or backing up the server, as well as the server's health state.

disableAutomatedBackup :: Maybe Bool

Disables automated backups. The number of stored backups is dependent on the value of PreferredBackupCount.

engineAttributes :: Maybe [EngineAttribute]

The response of a createServer() request returns the master credential to access the server in EngineAttributes. These credentials are not stored by AWS OpsWorks CM; they are returned only as part of the result of createServer().

Attributes returned in a createServer response for Chef

CHEF_AUTOMATE_PIVOTAL_KEY: A base64-encoded RSA private key that is generated by AWS OpsWorks for Chef Automate. This private key is required to access the Chef API.

CHEF_STARTER_KIT: A base64-encoded ZIP file. The ZIP file contains a Chef starter kit, which includes a README, a configuration file, and the required RSA private key. Save this file, unzip it, and then change to the directory where you've unzipped the file contents. From this directory, you can run Knife commands.

Attributes returned in a createServer response for Puppet

PUPPET_STARTER_KIT: A base64-encoded ZIP file. The ZIP file contains a Puppet starter kit, including a README and a required private key. Save this file, unzip it, and then change to the directory where you've unzipped the file contents.

PUPPET_ADMIN_PASSWORD: An administrator password that you can use to sign in to the Puppet Enterprise console after the server is online.

instanceType :: Maybe Text

The instance type for the server, as specified in the CloudFormation stack. This might not be the same instance type that is shown in the EC2 console.

customDomain :: Maybe Text

An optional public endpoint of a server, such as https://aws.my-company.com. You cannot access the server by using the Endpoint value if the server has a CustomDomain specified.

createdAt :: Maybe POSIX

Time stamp of server creation. Example 2016-07-29T13:38:47.520Z

subnetIds :: Maybe [Text]

The subnet IDs specified in a CreateServer request.

keyPair :: Maybe Text

The key pair associated with the server.

associatePublicIpAddress :: Maybe Bool

Associate a public IP address with a server that you are launching.

serverName :: Maybe Text

The name of the server.

engineVersion :: Maybe Text

The engine version of the server. For a Chef server, the valid value for EngineVersion is currently 2. For a Puppet server, specify either 2019 or 2017.

preferredMaintenanceWindow :: Maybe Text

The preferred maintenance period specified for the server.

cloudFormationStackArn :: Maybe Text

The ARN of the CloudFormation stack that was used to create the server.

maintenanceStatus :: Maybe MaintenanceStatus

The status of the most recent server maintenance run. Shows SUCCESS or FAILED.

engineModel :: Maybe Text

The engine model of the server. Valid values in this release include Monolithic for Puppet and Single for Chef.

backupRetentionCount :: Maybe Int

The number of automated backups to keep.

engine :: Maybe Text

The engine type of the server. Valid values in this release include ChefAutomate and Puppet.

endpoint :: Maybe Text

A DNS name that can be used to access the engine. Example: myserver-asdfghjkl.us-east-1.opsworks.io. You cannot access the server by using the Endpoint value if the server has a CustomDomain specified.

statusReason :: Maybe Text

Depending on the server status, this field has either a human-readable message (such as a create or backup error), or an escaped block of JSON (used for health check results).

serverArn :: Maybe Text

The ARN of the server.

newServer :: Server

Create a value of Server with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceProfileArn:Server', server_instanceProfileArn - The instance profile ARN of the server.

$sel:preferredBackupWindow:Server', server_preferredBackupWindow - The preferred backup period specified for the server.

$sel:securityGroupIds:Server', server_securityGroupIds - The security group IDs for the server, as specified in the CloudFormation stack. These might not be the same security groups that are shown in the EC2 console.

$sel:serviceRoleArn:Server', server_serviceRoleArn - The service role ARN used to create the server.

$sel:status:Server', server_status - The server's status. This field displays the states of actions in progress, such as creating, running, or backing up the server, as well as the server's health state.

$sel:disableAutomatedBackup:Server', server_disableAutomatedBackup - Disables automated backups. The number of stored backups is dependent on the value of PreferredBackupCount.

$sel:engineAttributes:Server', server_engineAttributes - The response of a createServer() request returns the master credential to access the server in EngineAttributes. These credentials are not stored by AWS OpsWorks CM; they are returned only as part of the result of createServer().

Attributes returned in a createServer response for Chef

CHEF_AUTOMATE_PIVOTAL_KEY: A base64-encoded RSA private key that is generated by AWS OpsWorks for Chef Automate. This private key is required to access the Chef API.

CHEF_STARTER_KIT: A base64-encoded ZIP file. The ZIP file contains a Chef starter kit, which includes a README, a configuration file, and the required RSA private key. Save this file, unzip it, and then change to the directory where you've unzipped the file contents. From this directory, you can run Knife commands.

Attributes returned in a createServer response for Puppet

PUPPET_STARTER_KIT: A base64-encoded ZIP file. The ZIP file contains a Puppet starter kit, including a README and a required private key. Save this file, unzip it, and then change to the directory where you've unzipped the file contents.

PUPPET_ADMIN_PASSWORD: An administrator password that you can use to sign in to the Puppet Enterprise console after the server is online.

$sel:instanceType:Server', server_instanceType - The instance type for the server, as specified in the CloudFormation stack. This might not be the same instance type that is shown in the EC2 console.

$sel:customDomain:Server', server_customDomain - An optional public endpoint of a server, such as https://aws.my-company.com. You cannot access the server by using the Endpoint value if the server has a CustomDomain specified.

$sel:createdAt:Server', server_createdAt - Time stamp of server creation. Example 2016-07-29T13:38:47.520Z

$sel:subnetIds:Server', server_subnetIds - The subnet IDs specified in a CreateServer request.

$sel:keyPair:Server', server_keyPair - The key pair associated with the server.

$sel:associatePublicIpAddress:Server', server_associatePublicIpAddress - Associate a public IP address with a server that you are launching.

$sel:serverName:Server', server_serverName - The name of the server.

$sel:engineVersion:Server', server_engineVersion - The engine version of the server. For a Chef server, the valid value for EngineVersion is currently 2. For a Puppet server, specify either 2019 or 2017.

$sel:preferredMaintenanceWindow:Server', server_preferredMaintenanceWindow - The preferred maintenance period specified for the server.

$sel:cloudFormationStackArn:Server', server_cloudFormationStackArn - The ARN of the CloudFormation stack that was used to create the server.

$sel:maintenanceStatus:Server', server_maintenanceStatus - The status of the most recent server maintenance run. Shows SUCCESS or FAILED.

$sel:engineModel:Server', server_engineModel - The engine model of the server. Valid values in this release include Monolithic for Puppet and Single for Chef.

$sel:backupRetentionCount:Server', server_backupRetentionCount - The number of automated backups to keep.

$sel:engine:Server', server_engine - The engine type of the server. Valid values in this release include ChefAutomate and Puppet.

$sel:endpoint:Server', server_endpoint - A DNS name that can be used to access the engine. Example: myserver-asdfghjkl.us-east-1.opsworks.io. You cannot access the server by using the Endpoint value if the server has a CustomDomain specified.

$sel:statusReason:Server', server_statusReason - Depending on the server status, this field has either a human-readable message (such as a create or backup error), or an escaped block of JSON (used for health check results).

$sel:serverArn:Server', server_serverArn - The ARN of the server.

server_instanceProfileArn :: Lens' Server (Maybe Text)

The instance profile ARN of the server.

server_preferredBackupWindow :: Lens' Server (Maybe Text)

The preferred backup period specified for the server.

server_securityGroupIds :: Lens' Server (Maybe [Text])

The security group IDs for the server, as specified in the CloudFormation stack. These might not be the same security groups that are shown in the EC2 console.

server_serviceRoleArn :: Lens' Server (Maybe Text)

The service role ARN used to create the server.

server_status :: Lens' Server (Maybe ServerStatus)

The server's status. This field displays the states of actions in progress, such as creating, running, or backing up the server, as well as the server's health state.

server_disableAutomatedBackup :: Lens' Server (Maybe Bool)

Disables automated backups. The number of stored backups is dependent on the value of PreferredBackupCount.

server_engineAttributes :: Lens' Server (Maybe [EngineAttribute])

The response of a createServer() request returns the master credential to access the server in EngineAttributes. These credentials are not stored by AWS OpsWorks CM; they are returned only as part of the result of createServer().

Attributes returned in a createServer response for Chef

CHEF_AUTOMATE_PIVOTAL_KEY: A base64-encoded RSA private key that is generated by AWS OpsWorks for Chef Automate. This private key is required to access the Chef API.

CHEF_STARTER_KIT: A base64-encoded ZIP file. The ZIP file contains a Chef starter kit, which includes a README, a configuration file, and the required RSA private key. Save this file, unzip it, and then change to the directory where you've unzipped the file contents. From this directory, you can run Knife commands.

Attributes returned in a createServer response for Puppet

PUPPET_STARTER_KIT: A base64-encoded ZIP file. The ZIP file contains a Puppet starter kit, including a README and a required private key. Save this file, unzip it, and then change to the directory where you've unzipped the file contents.

PUPPET_ADMIN_PASSWORD: An administrator password that you can use to sign in to the Puppet Enterprise console after the server is online.

server_instanceType :: Lens' Server (Maybe Text)

The instance type for the server, as specified in the CloudFormation stack. This might not be the same instance type that is shown in the EC2 console.

server_customDomain :: Lens' Server (Maybe Text)

An optional public endpoint of a server, such as https://aws.my-company.com. You cannot access the server by using the Endpoint value if the server has a CustomDomain specified.

server_createdAt :: Lens' Server (Maybe UTCTime)

Time stamp of server creation. Example 2016-07-29T13:38:47.520Z

server_subnetIds :: Lens' Server (Maybe [Text])

The subnet IDs specified in a CreateServer request.

server_keyPair :: Lens' Server (Maybe Text)

The key pair associated with the server.

server_associatePublicIpAddress :: Lens' Server (Maybe Bool)

Associate a public IP address with a server that you are launching.

server_serverName :: Lens' Server (Maybe Text)

The name of the server.

server_engineVersion :: Lens' Server (Maybe Text)

The engine version of the server. For a Chef server, the valid value for EngineVersion is currently 2. For a Puppet server, specify either 2019 or 2017.

server_preferredMaintenanceWindow :: Lens' Server (Maybe Text)

The preferred maintenance period specified for the server.

server_cloudFormationStackArn :: Lens' Server (Maybe Text)

The ARN of the CloudFormation stack that was used to create the server.

server_maintenanceStatus :: Lens' Server (Maybe MaintenanceStatus)

The status of the most recent server maintenance run. Shows SUCCESS or FAILED.

server_engineModel :: Lens' Server (Maybe Text)

The engine model of the server. Valid values in this release include Monolithic for Puppet and Single for Chef.

server_backupRetentionCount :: Lens' Server (Maybe Int)

The number of automated backups to keep.

server_engine :: Lens' Server (Maybe Text)

The engine type of the server. Valid values in this release include ChefAutomate and Puppet.

server_endpoint :: Lens' Server (Maybe Text)

A DNS name that can be used to access the engine. Example: myserver-asdfghjkl.us-east-1.opsworks.io. You cannot access the server by using the Endpoint value if the server has a CustomDomain specified.

server_statusReason :: Lens' Server (Maybe Text)

Depending on the server status, this field has either a human-readable message (such as a create or backup error), or an escaped block of JSON (used for health check results).

server_serverArn :: Lens' Server (Maybe Text)

The ARN of the server.