Network.AWS.Organizations.CreateOrganization

Description

Creates an AWS organization. The account whose user is calling the CreateOrganization operation automatically becomes the management account of the new organization.

This operation must be called using credentials from the account that is to become the new organization's management account. The principal must also have the relevant IAM permissions.

By default (or if you set the FeatureSet parameter to ALL), the new organization is created with all features enabled and service control policies automatically enabled in the root. If you instead choose to create the organization supporting only the consolidated billing features by setting the FeatureSet parameter to CONSOLIDATED_BILLING", no policy types are enabled by default, and you can't use organization policies

Creating a Request

data CreateOrganization

See: newCreateOrganization smart constructor.

Constructors

CreateOrganization'

Fields

featureSet :: Maybe OrganizationFeatureSet

Specifies the feature set supported by the new organization. Each feature set supports different levels of functionality.

CONSOLIDATED_BILLING: All member accounts have their bills consolidated to and paid by the management account. For more information, see Consolidated billing in the AWS Organizations User Guide.

The consolidated billing feature subset isn't available for organizations in the AWS GovCloud (US) Region.

ALL: In addition to all the features supported by the consolidated billing feature set, the management account can also apply any policy type to any member account in the organization. For more information, see All features in the AWS Organizations User Guide.

newCreateOrganization :: CreateOrganization

Create a value of CreateOrganization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:featureSet:CreateOrganization', createOrganization_featureSet - Specifies the feature set supported by the new organization. Each feature set supports different levels of functionality.

CONSOLIDATED_BILLING: All member accounts have their bills consolidated to and paid by the management account. For more information, see Consolidated billing in the AWS Organizations User Guide.

The consolidated billing feature subset isn't available for organizations in the AWS GovCloud (US) Region.

ALL: In addition to all the features supported by the consolidated billing feature set, the management account can also apply any policy type to any member account in the organization. For more information, see All features in the AWS Organizations User Guide.

Request Lenses

createOrganization_featureSet :: Lens' CreateOrganization (Maybe OrganizationFeatureSet)

Specifies the feature set supported by the new organization. Each feature set supports different levels of functionality.

CONSOLIDATED_BILLING: All member accounts have their bills consolidated to and paid by the management account. For more information, see Consolidated billing in the AWS Organizations User Guide.

The consolidated billing feature subset isn't available for organizations in the AWS GovCloud (US) Region.

ALL: In addition to all the features supported by the consolidated billing feature set, the management account can also apply any policy type to any member account in the organization. For more information, see All features in the AWS Organizations User Guide.

Destructuring the Response

data CreateOrganizationResponse

See: newCreateOrganizationResponse smart constructor.

Constructors

CreateOrganizationResponse'

Fields

organization :: Maybe Organization

A structure that contains details about the newly created organization.

httpStatus :: Int

The response's http status code.

newCreateOrganizationResponse

Arguments

:: Int
$sel:httpStatus:CreateOrganizationResponse'

-> CreateOrganizationResponse

Create a value of CreateOrganizationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:organization:CreateOrganizationResponse', createOrganizationResponse_organization - A structure that contains details about the newly created organization.

$sel:httpStatus:CreateOrganizationResponse', createOrganizationResponse_httpStatus - The response's http status code.

Response Lenses

createOrganizationResponse_organization :: Lens' CreateOrganizationResponse (Maybe Organization)

A structure that contains details about the newly created organization.

createOrganizationResponse_httpStatus :: Lens' CreateOrganizationResponse Int

The response's http status code.