Network.AWS.Organizations.DeregisterDelegatedAdministrator

Description

Removes the specified member AWS account as a delegated administrator for the specified AWS service.

Deregistering a delegated administrator can have unintended impacts on the functionality of the enabled AWS service. See the documentation for the enabled service before you deregister a delegated administrator so that you understand any potential impacts.

You can run this action only for AWS services that support this feature. For a current list of services that support it, see the column /Supports Delegated Administrator/ in the table at AWS Services that you can use with AWS Organizations in the AWS Organizations User Guide.

This operation can be called only from the organization's management account.

Creating a Request

data DeregisterDelegatedAdministrator

See: newDeregisterDelegatedAdministrator smart constructor.

Constructors

DeregisterDelegatedAdministrator'

Fields

accountId :: Text

The account ID number of the member account in the organization that you want to deregister as a delegated administrator.

servicePrincipal :: Text

The service principal name of an AWS service for which the account is a delegated administrator.

Delegated administrator privileges are revoked for only the specified AWS service from the member account. If the specified service is the only service for which the member account is a delegated administrator, the operation also revokes Organizations read action permissions.

newDeregisterDelegatedAdministrator

Arguments

:: Text
$sel:accountId:DeregisterDelegatedAdministrator'

-> Text
$sel:servicePrincipal:DeregisterDelegatedAdministrator'

-> DeregisterDelegatedAdministrator

Create a value of DeregisterDelegatedAdministrator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:DeregisterDelegatedAdministrator', deregisterDelegatedAdministrator_accountId - The account ID number of the member account in the organization that you want to deregister as a delegated administrator.

$sel:servicePrincipal:DeregisterDelegatedAdministrator', deregisterDelegatedAdministrator_servicePrincipal - The service principal name of an AWS service for which the account is a delegated administrator.

Delegated administrator privileges are revoked for only the specified AWS service from the member account. If the specified service is the only service for which the member account is a delegated administrator, the operation also revokes Organizations read action permissions.

Request Lenses

deregisterDelegatedAdministrator_accountId :: Lens' DeregisterDelegatedAdministrator Text

The account ID number of the member account in the organization that you want to deregister as a delegated administrator.

deregisterDelegatedAdministrator_servicePrincipal :: Lens' DeregisterDelegatedAdministrator Text

The service principal name of an AWS service for which the account is a delegated administrator.

Delegated administrator privileges are revoked for only the specified AWS service from the member account. If the specified service is the only service for which the member account is a delegated administrator, the operation also revokes Organizations read action permissions.

Destructuring the Response

data DeregisterDelegatedAdministratorResponse

See: newDeregisterDelegatedAdministratorResponse smart constructor.

Constructors

DeregisterDelegatedAdministratorResponse'

newDeregisterDelegatedAdministratorResponse :: DeregisterDelegatedAdministratorResponse

Create a value of DeregisterDelegatedAdministratorResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.