Network.AWS.Organizations.DescribeHandshake

Description

Retrieves information about a previously requested handshake. The handshake ID comes from the response to the original InviteAccountToOrganization operation that generated the handshake.

You can access handshakes that are ACCEPTED, DECLINED, or CANCELED for only 30 days after they change to that state. They're then deleted and no longer accessible.

This operation can be called from any account in the organization.

Creating a Request

data DescribeHandshake

See: newDescribeHandshake smart constructor.

Constructors

DescribeHandshake'

Fields

handshakeId :: Text

The unique identifier (ID) of the handshake that you want information about. You can get the ID from the original call to InviteAccountToOrganization, or from a call to ListHandshakesForAccount or ListHandshakesForOrganization.

The regex pattern for handshake ID string requires "h-" followed by from 8 to 32 lowercase letters or digits.

newDescribeHandshake

Arguments

:: Text
$sel:handshakeId:DescribeHandshake'

-> DescribeHandshake

Create a value of DescribeHandshake with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:handshakeId:DescribeHandshake', describeHandshake_handshakeId - The unique identifier (ID) of the handshake that you want information about. You can get the ID from the original call to InviteAccountToOrganization, or from a call to ListHandshakesForAccount or ListHandshakesForOrganization.

The regex pattern for handshake ID string requires "h-" followed by from 8 to 32 lowercase letters or digits.

Request Lenses

describeHandshake_handshakeId :: Lens' DescribeHandshake Text

The unique identifier (ID) of the handshake that you want information about. You can get the ID from the original call to InviteAccountToOrganization, or from a call to ListHandshakesForAccount or ListHandshakesForOrganization.

The regex pattern for handshake ID string requires "h-" followed by from 8 to 32 lowercase letters or digits.

Destructuring the Response

data DescribeHandshakeResponse

See: newDescribeHandshakeResponse smart constructor.

Constructors

DescribeHandshakeResponse'

Fields

handshake :: Maybe Handshake

A structure that contains information about the specified handshake.

httpStatus :: Int

The response's http status code.

newDescribeHandshakeResponse

Arguments

:: Int
$sel:httpStatus:DescribeHandshakeResponse'

-> DescribeHandshakeResponse

Create a value of DescribeHandshakeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:handshake:DescribeHandshakeResponse', describeHandshakeResponse_handshake - A structure that contains information about the specified handshake.

$sel:httpStatus:DescribeHandshakeResponse', describeHandshakeResponse_httpStatus - The response's http status code.

Response Lenses

describeHandshakeResponse_handshake :: Lens' DescribeHandshakeResponse (Maybe Handshake)

A structure that contains information about the specified handshake.

describeHandshakeResponse_httpStatus :: Lens' DescribeHandshakeResponse Int

The response's http status code.