Network.AWS.Organizations.ListAccounts

Description

Lists all the accounts in the organization. To request only the accounts in a specified root or organizational unit (OU), use the ListAccountsForParent operation instead.

Always check the NextToken response parameter for a null value when calling a List* operation. These operations can occasionally return an empty set of results even when there are more results available. The NextToken response parameter value is null only when there are no more results to display.

This operation can be called only from the organization's management account or by a member account that is a delegated administrator for an AWS service.

This operation returns paginated results.

Creating a Request

data ListAccounts

See: newListAccounts smart constructor.

Constructors

ListAccounts'

Fields

nextToken :: Maybe Text

The parameter for receiving additional results if you receive a NextToken response in a previous request. A NextToken response indicates that more output is available. Set this parameter to the value of the previous call's NextToken response to indicate where the output should continue from.

maxResults :: Maybe Natural

The total number of results that you want included on each page of the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

newListAccounts :: ListAccounts

Create a value of ListAccounts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAccounts', listAccounts_nextToken - The parameter for receiving additional results if you receive a NextToken response in a previous request. A NextToken response indicates that more output is available. Set this parameter to the value of the previous call's NextToken response to indicate where the output should continue from.

$sel:maxResults:ListAccounts', listAccounts_maxResults - The total number of results that you want included on each page of the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

Request Lenses

listAccounts_nextToken :: Lens' ListAccounts (Maybe Text)

The parameter for receiving additional results if you receive a NextToken response in a previous request. A NextToken response indicates that more output is available. Set this parameter to the value of the previous call's NextToken response to indicate where the output should continue from.

listAccounts_maxResults :: Lens' ListAccounts (Maybe Natural)

The total number of results that you want included on each page of the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

Destructuring the Response

data ListAccountsResponse

See: newListAccountsResponse smart constructor.

Constructors

ListAccountsResponse'

Fields

nextToken :: Maybe Text

If present, indicates that more output is available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null.

accounts :: Maybe [Account]

A list of objects in the organization.

httpStatus :: Int

The response's http status code.

newListAccountsResponse

Arguments

:: Int
$sel:httpStatus:ListAccountsResponse'

-> ListAccountsResponse

Create a value of ListAccountsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAccounts', listAccountsResponse_nextToken - If present, indicates that more output is available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null.

$sel:accounts:ListAccountsResponse', listAccountsResponse_accounts - A list of objects in the organization.

$sel:httpStatus:ListAccountsResponse', listAccountsResponse_httpStatus - The response's http status code.

Response Lenses

listAccountsResponse_nextToken :: Lens' ListAccountsResponse (Maybe Text)

If present, indicates that more output is available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null.

listAccountsResponse_accounts :: Lens' ListAccountsResponse (Maybe [Account])

A list of objects in the organization.

listAccountsResponse_httpStatus :: Lens' ListAccountsResponse Int

The response's http status code.