Network.AWS.Organizations.ListAccountsForParent

Description

Lists the accounts in an organization that are contained by the specified target root or organizational unit (OU). If you specify the root, you get a list of all the accounts that aren't in any OU. If you specify an OU, you get a list of all the accounts in only that OU and not in any child OUs. To get a list of all accounts in the organization, use the ListAccounts operation.

Always check the NextToken response parameter for a null value when calling a List* operation. These operations can occasionally return an empty set of results even when there are more results available. The NextToken response parameter value is null only when there are no more results to display.

This operation can be called only from the organization's management account or by a member account that is a delegated administrator for an AWS service.

This operation returns paginated results.

Creating a Request

data ListAccountsForParent

See: newListAccountsForParent smart constructor.

Constructors

ListAccountsForParent'

Fields

nextToken :: Maybe Text

The parameter for receiving additional results if you receive a NextToken response in a previous request. A NextToken response indicates that more output is available. Set this parameter to the value of the previous call's NextToken response to indicate where the output should continue from.

maxResults :: Maybe Natural

The total number of results that you want included on each page of the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

parentId :: Text

The unique identifier (ID) for the parent root or organization unit (OU) whose accounts you want to list.

newListAccountsForParent

Arguments

:: Text
$sel:parentId:ListAccountsForParent'

-> ListAccountsForParent

Create a value of ListAccountsForParent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAccountsForParent', listAccountsForParent_nextToken - The parameter for receiving additional results if you receive a NextToken response in a previous request. A NextToken response indicates that more output is available. Set this parameter to the value of the previous call's NextToken response to indicate where the output should continue from.

$sel:maxResults:ListAccountsForParent', listAccountsForParent_maxResults - The total number of results that you want included on each page of the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

$sel:parentId:ListAccountsForParent', listAccountsForParent_parentId - The unique identifier (ID) for the parent root or organization unit (OU) whose accounts you want to list.

Request Lenses

listAccountsForParent_nextToken :: Lens' ListAccountsForParent (Maybe Text)

The parameter for receiving additional results if you receive a NextToken response in a previous request. A NextToken response indicates that more output is available. Set this parameter to the value of the previous call's NextToken response to indicate where the output should continue from.

listAccountsForParent_maxResults :: Lens' ListAccountsForParent (Maybe Natural)

The total number of results that you want included on each page of the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

listAccountsForParent_parentId :: Lens' ListAccountsForParent Text

The unique identifier (ID) for the parent root or organization unit (OU) whose accounts you want to list.

Destructuring the Response

data ListAccountsForParentResponse

See: newListAccountsForParentResponse smart constructor.

Constructors

ListAccountsForParentResponse'

Fields

nextToken :: Maybe Text

If present, indicates that more output is available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null.

accounts :: Maybe [Account]

A list of the accounts in the specified root or OU.

httpStatus :: Int

The response's http status code.

newListAccountsForParentResponse

Arguments

:: Int
$sel:httpStatus:ListAccountsForParentResponse'

-> ListAccountsForParentResponse

Create a value of ListAccountsForParentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListAccountsForParent', listAccountsForParentResponse_nextToken - If present, indicates that more output is available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null.

$sel:accounts:ListAccountsForParentResponse', listAccountsForParentResponse_accounts - A list of the accounts in the specified root or OU.

$sel:httpStatus:ListAccountsForParentResponse', listAccountsForParentResponse_httpStatus - The response's http status code.

Response Lenses

listAccountsForParentResponse_nextToken :: Lens' ListAccountsForParentResponse (Maybe Text)

If present, indicates that more output is available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null.

listAccountsForParentResponse_accounts :: Lens' ListAccountsForParentResponse (Maybe [Account])

A list of the accounts in the specified root or OU.

listAccountsForParentResponse_httpStatus :: Lens' ListAccountsForParentResponse Int

The response's http status code.