Network.AWS.Organizations.ListCreateAccountStatus

Description

Lists the account creation requests that match the specified status that is currently being tracked for the organization.

Always check the NextToken response parameter for a null value when calling a List* operation. These operations can occasionally return an empty set of results even when there are more results available. The NextToken response parameter value is null only when there are no more results to display.

This operation can be called only from the organization's management account or by a member account that is a delegated administrator for an AWS service.

This operation returns paginated results.

Creating a Request

data ListCreateAccountStatus

See: newListCreateAccountStatus smart constructor.

Constructors

ListCreateAccountStatus'

Fields

states :: Maybe [CreateAccountState]

A list of one or more states that you want included in the response. If this parameter isn't present, all requests are included in the response.

nextToken :: Maybe Text

The parameter for receiving additional results if you receive a NextToken response in a previous request. A NextToken response indicates that more output is available. Set this parameter to the value of the previous call's NextToken response to indicate where the output should continue from.

maxResults :: Maybe Natural

The total number of results that you want included on each page of the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

newListCreateAccountStatus :: ListCreateAccountStatus

Create a value of ListCreateAccountStatus with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:states:ListCreateAccountStatus', listCreateAccountStatus_states - A list of one or more states that you want included in the response. If this parameter isn't present, all requests are included in the response.

$sel:nextToken:ListCreateAccountStatus', listCreateAccountStatus_nextToken - The parameter for receiving additional results if you receive a NextToken response in a previous request. A NextToken response indicates that more output is available. Set this parameter to the value of the previous call's NextToken response to indicate where the output should continue from.

$sel:maxResults:ListCreateAccountStatus', listCreateAccountStatus_maxResults - The total number of results that you want included on each page of the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

Request Lenses

listCreateAccountStatus_states :: Lens' ListCreateAccountStatus (Maybe [CreateAccountState])

A list of one or more states that you want included in the response. If this parameter isn't present, all requests are included in the response.

listCreateAccountStatus_nextToken :: Lens' ListCreateAccountStatus (Maybe Text)

The parameter for receiving additional results if you receive a NextToken response in a previous request. A NextToken response indicates that more output is available. Set this parameter to the value of the previous call's NextToken response to indicate where the output should continue from.

listCreateAccountStatus_maxResults :: Lens' ListCreateAccountStatus (Maybe Natural)

The total number of results that you want included on each page of the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

Destructuring the Response

data ListCreateAccountStatusResponse

See: newListCreateAccountStatusResponse smart constructor.

Constructors

ListCreateAccountStatusResponse'

Fields

nextToken :: Maybe Text

If present, indicates that more output is available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null.

createAccountStatuses :: Maybe [CreateAccountStatus]

A list of objects with details about the requests. Certain elements, such as the accountId number, are present in the output only after the account has been successfully created.

httpStatus :: Int

The response's http status code.

newListCreateAccountStatusResponse

Arguments

:: Int
$sel:httpStatus:ListCreateAccountStatusResponse'

-> ListCreateAccountStatusResponse

Create a value of ListCreateAccountStatusResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListCreateAccountStatus', listCreateAccountStatusResponse_nextToken - If present, indicates that more output is available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null.

$sel:createAccountStatuses:ListCreateAccountStatusResponse', listCreateAccountStatusResponse_createAccountStatuses - A list of objects with details about the requests. Certain elements, such as the accountId number, are present in the output only after the account has been successfully created.

$sel:httpStatus:ListCreateAccountStatusResponse', listCreateAccountStatusResponse_httpStatus - The response's http status code.

Response Lenses

listCreateAccountStatusResponse_nextToken :: Lens' ListCreateAccountStatusResponse (Maybe Text)

If present, indicates that more output is available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null.

listCreateAccountStatusResponse_createAccountStatuses :: Lens' ListCreateAccountStatusResponse (Maybe [CreateAccountStatus])

A list of objects with details about the requests. Certain elements, such as the accountId number, are present in the output only after the account has been successfully created.

listCreateAccountStatusResponse_httpStatus :: Lens' ListCreateAccountStatusResponse Int

The response's http status code.