Network.AWS.Organizations.ListDelegatedAdministrators

Description

Lists the AWS accounts that are designated as delegated administrators in this organization.

This operation can be called only from the organization's management account or by a member account that is a delegated administrator for an AWS service.

This operation returns paginated results.

Creating a Request

data ListDelegatedAdministrators

See: newListDelegatedAdministrators smart constructor.

Constructors

ListDelegatedAdministrators'

Fields

servicePrincipal :: Maybe Text

Specifies a service principal name. If specified, then the operation lists the delegated administrators only for the specified service.

If you don't specify a service principal, the operation lists all delegated administrators for all services in your organization.

nextToken :: Maybe Text

The parameter for receiving additional results if you receive a NextToken response in a previous request. A NextToken response indicates that more output is available. Set this parameter to the value of the previous call's NextToken response to indicate where the output should continue from.

maxResults :: Maybe Natural

The total number of results that you want included on each page of the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

newListDelegatedAdministrators :: ListDelegatedAdministrators

Create a value of ListDelegatedAdministrators with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:servicePrincipal:ListDelegatedAdministrators', listDelegatedAdministrators_servicePrincipal - Specifies a service principal name. If specified, then the operation lists the delegated administrators only for the specified service.

If you don't specify a service principal, the operation lists all delegated administrators for all services in your organization.

$sel:nextToken:ListDelegatedAdministrators', listDelegatedAdministrators_nextToken - The parameter for receiving additional results if you receive a NextToken response in a previous request. A NextToken response indicates that more output is available. Set this parameter to the value of the previous call's NextToken response to indicate where the output should continue from.

$sel:maxResults:ListDelegatedAdministrators', listDelegatedAdministrators_maxResults - The total number of results that you want included on each page of the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

Request Lenses

listDelegatedAdministrators_servicePrincipal :: Lens' ListDelegatedAdministrators (Maybe Text)

Specifies a service principal name. If specified, then the operation lists the delegated administrators only for the specified service.

If you don't specify a service principal, the operation lists all delegated administrators for all services in your organization.

listDelegatedAdministrators_nextToken :: Lens' ListDelegatedAdministrators (Maybe Text)

The parameter for receiving additional results if you receive a NextToken response in a previous request. A NextToken response indicates that more output is available. Set this parameter to the value of the previous call's NextToken response to indicate where the output should continue from.

listDelegatedAdministrators_maxResults :: Lens' ListDelegatedAdministrators (Maybe Natural)

The total number of results that you want included on each page of the response. If you do not include this parameter, it defaults to a value that is specific to the operation. If additional items exist beyond the maximum you specify, the NextToken response element is present and has a value (is not null). Include that value as the NextToken request parameter in the next call to the operation to get the next part of the results. Note that Organizations might return fewer results than the maximum even when there are more results available. You should check NextToken after every operation to ensure that you receive all of the results.

Destructuring the Response

data ListDelegatedAdministratorsResponse

See: newListDelegatedAdministratorsResponse smart constructor.

Constructors

ListDelegatedAdministratorsResponse'

Fields

delegatedAdministrators :: Maybe [DelegatedAdministrator]

The list of delegated administrators in your organization.

nextToken :: Maybe Text

If present, indicates that more output is available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null.

httpStatus :: Int

The response's http status code.

newListDelegatedAdministratorsResponse

Arguments

:: Int
$sel:httpStatus:ListDelegatedAdministratorsResponse'

-> ListDelegatedAdministratorsResponse

Create a value of ListDelegatedAdministratorsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:delegatedAdministrators:ListDelegatedAdministratorsResponse', listDelegatedAdministratorsResponse_delegatedAdministrators - The list of delegated administrators in your organization.

$sel:nextToken:ListDelegatedAdministrators', listDelegatedAdministratorsResponse_nextToken - If present, indicates that more output is available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null.

$sel:httpStatus:ListDelegatedAdministratorsResponse', listDelegatedAdministratorsResponse_httpStatus - The response's http status code.

Response Lenses

listDelegatedAdministratorsResponse_delegatedAdministrators :: Lens' ListDelegatedAdministratorsResponse (Maybe [DelegatedAdministrator])

The list of delegated administrators in your organization.

listDelegatedAdministratorsResponse_nextToken :: Lens' ListDelegatedAdministratorsResponse (Maybe Text)

If present, indicates that more output is available than is included in the current response. Use this value in the NextToken request parameter in a subsequent call to the operation to get the next part of the output. You should repeat this until the NextToken response element comes back as null.

listDelegatedAdministratorsResponse_httpStatus :: Lens' ListDelegatedAdministratorsResponse Int

The response's http status code.