Network.AWS.Organizations.Types.DelegatedAdministrator

Description

Documentation

data DelegatedAdministrator

Contains information about the delegated administrator.

See: newDelegatedAdministrator smart constructor.

Constructors

DelegatedAdministrator'

Fields

status :: Maybe AccountStatus

The status of the delegated administrator's account in the organization.

joinedMethod :: Maybe AccountJoinedMethod

The method by which the delegated administrator's account joined the organization.

email :: Maybe (Sensitive Text)

The email address that is associated with the delegated administrator's AWS account.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the delegated administrator's account.

joinedTimestamp :: Maybe POSIX

The date when the delegated administrator's account became a part of the organization.

delegationEnabledDate :: Maybe POSIX

The date when the account was made a delegated administrator.

name :: Maybe (Sensitive Text)

The friendly name of the delegated administrator's account.

id :: Maybe Text

The unique identifier (ID) of the delegated administrator's account.

newDelegatedAdministrator :: DelegatedAdministrator

Create a value of DelegatedAdministrator with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DelegatedAdministrator', delegatedAdministrator_status - The status of the delegated administrator's account in the organization.

$sel:joinedMethod:DelegatedAdministrator', delegatedAdministrator_joinedMethod - The method by which the delegated administrator's account joined the organization.

$sel:email:DelegatedAdministrator', delegatedAdministrator_email - The email address that is associated with the delegated administrator's AWS account.

$sel:arn:DelegatedAdministrator', delegatedAdministrator_arn - The Amazon Resource Name (ARN) of the delegated administrator's account.

$sel:joinedTimestamp:DelegatedAdministrator', delegatedAdministrator_joinedTimestamp - The date when the delegated administrator's account became a part of the organization.

$sel:delegationEnabledDate:DelegatedAdministrator', delegatedAdministrator_delegationEnabledDate - The date when the account was made a delegated administrator.

$sel:name:DelegatedAdministrator', delegatedAdministrator_name - The friendly name of the delegated administrator's account.

$sel:id:DelegatedAdministrator', delegatedAdministrator_id - The unique identifier (ID) of the delegated administrator's account.

delegatedAdministrator_status :: Lens' DelegatedAdministrator (Maybe AccountStatus)

The status of the delegated administrator's account in the organization.

delegatedAdministrator_joinedMethod :: Lens' DelegatedAdministrator (Maybe AccountJoinedMethod)

The method by which the delegated administrator's account joined the organization.

delegatedAdministrator_email :: Lens' DelegatedAdministrator (Maybe Text)

The email address that is associated with the delegated administrator's AWS account.

delegatedAdministrator_arn :: Lens' DelegatedAdministrator (Maybe Text)

The Amazon Resource Name (ARN) of the delegated administrator's account.

delegatedAdministrator_joinedTimestamp :: Lens' DelegatedAdministrator (Maybe UTCTime)

The date when the delegated administrator's account became a part of the organization.

delegatedAdministrator_delegationEnabledDate :: Lens' DelegatedAdministrator (Maybe UTCTime)

The date when the account was made a delegated administrator.

delegatedAdministrator_name :: Lens' DelegatedAdministrator (Maybe Text)

The friendly name of the delegated administrator's account.

delegatedAdministrator_id :: Lens' DelegatedAdministrator (Maybe Text)

The unique identifier (ID) of the delegated administrator's account.