Network.AWS.Organizations.Types.EffectivePolicy

Description

Documentation

data EffectivePolicy

Contains rules to be applied to the affected accounts. The effective policy is the aggregation of any policies the account inherits, plus any policy directly attached to the account.

See: newEffectivePolicy smart constructor.

Constructors

EffectivePolicy'

Fields

targetId :: Maybe Text

The account ID of the policy target.

policyType :: Maybe EffectivePolicyType

The policy type.

lastUpdatedTimestamp :: Maybe POSIX

The time of the last update to this policy.

policyContent :: Maybe Text

The text content of the policy.

newEffectivePolicy :: EffectivePolicy

Create a value of EffectivePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetId:EffectivePolicy', effectivePolicy_targetId - The account ID of the policy target.

$sel:policyType:EffectivePolicy', effectivePolicy_policyType - The policy type.

$sel:lastUpdatedTimestamp:EffectivePolicy', effectivePolicy_lastUpdatedTimestamp - The time of the last update to this policy.

$sel:policyContent:EffectivePolicy', effectivePolicy_policyContent - The text content of the policy.

effectivePolicy_targetId :: Lens' EffectivePolicy (Maybe Text)

The account ID of the policy target.

effectivePolicy_policyType :: Lens' EffectivePolicy (Maybe EffectivePolicyType)

The policy type.

effectivePolicy_lastUpdatedTimestamp :: Lens' EffectivePolicy (Maybe UTCTime)

The time of the last update to this policy.

effectivePolicy_policyContent :: Lens' EffectivePolicy (Maybe Text)

The text content of the policy.