Network.AWS.Organizations.Types.Handshake

Description

Documentation

data Handshake

Contains information that must be exchanged to securely establish a relationship between two accounts (an originator and a recipient). For example, when a management account (the originator) invites another account (the recipient) to join its organization, the two accounts exchange information as a series of handshake requests and responses.

Note: Handshakes that are CANCELED, ACCEPTED, or DECLINED show up in lists for only 30 days after entering that state After that they are deleted.

See: newHandshake smart constructor.

Constructors

Handshake'

Fields

requestedTimestamp :: Maybe POSIX

The date and time that the handshake request was made.

parties :: Maybe [HandshakeParty]

Information about the two accounts that are participating in the handshake.

arn :: Maybe Text

The Amazon Resource Name (ARN) of a handshake.

For more information about ARNs in Organizations, see ARN Formats Supported by Organizations in the AWS Service Authorization Reference.

id :: Maybe Text

The unique identifier (ID) of a handshake. The originating account creates the ID when it initiates the handshake.

The regex pattern for handshake ID string requires "h-" followed by from 8 to 32 lowercase letters or digits.

state :: Maybe HandshakeState

The current state of the handshake. Use the state to trace the flow of the handshake through the process from its creation to its acceptance. The meaning of each of the valid values is as follows:

REQUESTED: This handshake was sent to multiple recipients (applicable to only some handshake types) and not all recipients have responded yet. The request stays in this state until all recipients respond.

OPEN: This handshake was sent to multiple recipients (applicable to only some policy types) and all recipients have responded, allowing the originator to complete the handshake action.

CANCELED: This handshake is no longer active because it was canceled by the originating account.

ACCEPTED: This handshake is complete because it has been accepted by the recipient.

DECLINED: This handshake is no longer active because it was declined by the recipient account.

EXPIRED: This handshake is no longer active because the originator did not receive a response of any kind from the recipient before the expiration time (15 days).

resources :: Maybe [HandshakeResource]

Additional information that is needed to process the handshake.

action :: Maybe ActionType

The type of handshake, indicating what action occurs when the recipient accepts the handshake. The following handshake types are supported:

INVITE: This type of handshake represents a request to join an organization. It is always sent from the management account to only non-member accounts.

ENABLE_ALL_FEATURES: This type of handshake represents a request to enable all features in an organization. It is always sent from the management account to only invited member accounts. Created accounts do not receive this because those accounts were created by the organization's management account and approval is inferred.

APPROVE_ALL_FEATURES: This type of handshake is sent from the Organizations service when all member accounts have approved the ENABLE_ALL_FEATURES invitation. It is sent only to the management account and signals the master that it can finalize the process to enable all features.

expirationTimestamp :: Maybe POSIX

The date and time that the handshake expires. If the recipient of the handshake request fails to respond before the specified date and time, the handshake becomes inactive and is no longer valid.

newHandshake :: Handshake

Create a value of Handshake with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestedTimestamp:Handshake', handshake_requestedTimestamp - The date and time that the handshake request was made.

$sel:parties:Handshake', handshake_parties - Information about the two accounts that are participating in the handshake.

$sel:arn:Handshake', handshake_arn - The Amazon Resource Name (ARN) of a handshake.

For more information about ARNs in Organizations, see ARN Formats Supported by Organizations in the AWS Service Authorization Reference.

$sel:id:Handshake', handshake_id - The unique identifier (ID) of a handshake. The originating account creates the ID when it initiates the handshake.

The regex pattern for handshake ID string requires "h-" followed by from 8 to 32 lowercase letters or digits.

$sel:state:Handshake', handshake_state - The current state of the handshake. Use the state to trace the flow of the handshake through the process from its creation to its acceptance. The meaning of each of the valid values is as follows:

REQUESTED: This handshake was sent to multiple recipients (applicable to only some handshake types) and not all recipients have responded yet. The request stays in this state until all recipients respond.

OPEN: This handshake was sent to multiple recipients (applicable to only some policy types) and all recipients have responded, allowing the originator to complete the handshake action.

CANCELED: This handshake is no longer active because it was canceled by the originating account.

ACCEPTED: This handshake is complete because it has been accepted by the recipient.

DECLINED: This handshake is no longer active because it was declined by the recipient account.

EXPIRED: This handshake is no longer active because the originator did not receive a response of any kind from the recipient before the expiration time (15 days).

$sel:resources:Handshake', handshake_resources - Additional information that is needed to process the handshake.

$sel:action:Handshake', handshake_action - The type of handshake, indicating what action occurs when the recipient accepts the handshake. The following handshake types are supported:

INVITE: This type of handshake represents a request to join an organization. It is always sent from the management account to only non-member accounts.

ENABLE_ALL_FEATURES: This type of handshake represents a request to enable all features in an organization. It is always sent from the management account to only invited member accounts. Created accounts do not receive this because those accounts were created by the organization's management account and approval is inferred.

APPROVE_ALL_FEATURES: This type of handshake is sent from the Organizations service when all member accounts have approved the ENABLE_ALL_FEATURES invitation. It is sent only to the management account and signals the master that it can finalize the process to enable all features.

$sel:expirationTimestamp:Handshake', handshake_expirationTimestamp - The date and time that the handshake expires. If the recipient of the handshake request fails to respond before the specified date and time, the handshake becomes inactive and is no longer valid.

handshake_requestedTimestamp :: Lens' Handshake (Maybe UTCTime)

The date and time that the handshake request was made.

handshake_parties :: Lens' Handshake (Maybe [HandshakeParty])

Information about the two accounts that are participating in the handshake.

handshake_arn :: Lens' Handshake (Maybe Text)

The Amazon Resource Name (ARN) of a handshake.

For more information about ARNs in Organizations, see ARN Formats Supported by Organizations in the AWS Service Authorization Reference.

handshake_id :: Lens' Handshake (Maybe Text)

The unique identifier (ID) of a handshake. The originating account creates the ID when it initiates the handshake.

The regex pattern for handshake ID string requires "h-" followed by from 8 to 32 lowercase letters or digits.

handshake_state :: Lens' Handshake (Maybe HandshakeState)

The current state of the handshake. Use the state to trace the flow of the handshake through the process from its creation to its acceptance. The meaning of each of the valid values is as follows:

REQUESTED: This handshake was sent to multiple recipients (applicable to only some handshake types) and not all recipients have responded yet. The request stays in this state until all recipients respond.

OPEN: This handshake was sent to multiple recipients (applicable to only some policy types) and all recipients have responded, allowing the originator to complete the handshake action.

CANCELED: This handshake is no longer active because it was canceled by the originating account.

ACCEPTED: This handshake is complete because it has been accepted by the recipient.

DECLINED: This handshake is no longer active because it was declined by the recipient account.

EXPIRED: This handshake is no longer active because the originator did not receive a response of any kind from the recipient before the expiration time (15 days).

handshake_resources :: Lens' Handshake (Maybe [HandshakeResource])

Additional information that is needed to process the handshake.

handshake_action :: Lens' Handshake (Maybe ActionType)

The type of handshake, indicating what action occurs when the recipient accepts the handshake. The following handshake types are supported:

INVITE: This type of handshake represents a request to join an organization. It is always sent from the management account to only non-member accounts.

ENABLE_ALL_FEATURES: This type of handshake represents a request to enable all features in an organization. It is always sent from the management account to only invited member accounts. Created accounts do not receive this because those accounts were created by the organization's management account and approval is inferred.

APPROVE_ALL_FEATURES: This type of handshake is sent from the Organizations service when all member accounts have approved the ENABLE_ALL_FEATURES invitation. It is sent only to the management account and signals the master that it can finalize the process to enable all features.

handshake_expirationTimestamp :: Lens' Handshake (Maybe UTCTime)

The date and time that the handshake expires. If the recipient of the handshake request fails to respond before the specified date and time, the handshake becomes inactive and is no longer valid.