Network.AWS.Organizations.Types.Organization

Description

Documentation

data Organization

Contains details about an organization. An organization is a collection of accounts that are centrally managed together using consolidated billing, organized hierarchically with organizational units (OUs), and controlled with policies .

See: newOrganization smart constructor.

Constructors

Organization'

Fields

masterAccountEmail :: Maybe (Sensitive Text)

The email address that is associated with the AWS account that is designated as the management account for the organization.

featureSet :: Maybe OrganizationFeatureSet

Specifies the functionality that currently is available to the organization. If set to "ALL", then all features are enabled and policies can be applied to accounts in the organization. If set to "CONSOLIDATED_BILLING", then only consolidated billing functionality is available. For more information, see Enabling All Features in Your Organization in the AWS Organizations User Guide.

masterAccountArn :: Maybe Text

The Amazon Resource Name (ARN) of the account that is designated as the management account for the organization.

For more information about ARNs in Organizations, see ARN Formats Supported by Organizations in the AWS Service Authorization Reference.

arn :: Maybe Text

The Amazon Resource Name (ARN) of an organization.

For more information about ARNs in Organizations, see ARN Formats Supported by Organizations in the AWS Service Authorization Reference.

id :: Maybe Text

The unique identifier (ID) of an organization.

The regex pattern for an organization ID string requires "o-" followed by from 10 to 32 lowercase letters or digits.

masterAccountId :: Maybe Text

The unique identifier (ID) of the management account of an organization.

The regex pattern for an account ID string requires exactly 12 digits.

availablePolicyTypes :: Maybe [PolicyTypeSummary]

Do not use. This field is deprecated and doesn't provide complete information about the policies in your organization.

To determine the policies that are enabled and available for use in your organization, use the ListRoots operation instead.

newOrganization :: Organization

Create a value of Organization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:masterAccountEmail:Organization', organization_masterAccountEmail - The email address that is associated with the AWS account that is designated as the management account for the organization.

$sel:featureSet:Organization', organization_featureSet - Specifies the functionality that currently is available to the organization. If set to "ALL", then all features are enabled and policies can be applied to accounts in the organization. If set to "CONSOLIDATED_BILLING", then only consolidated billing functionality is available. For more information, see Enabling All Features in Your Organization in the AWS Organizations User Guide.

$sel:masterAccountArn:Organization', organization_masterAccountArn - The Amazon Resource Name (ARN) of the account that is designated as the management account for the organization.

For more information about ARNs in Organizations, see ARN Formats Supported by Organizations in the AWS Service Authorization Reference.

$sel:arn:Organization', organization_arn - The Amazon Resource Name (ARN) of an organization.

For more information about ARNs in Organizations, see ARN Formats Supported by Organizations in the AWS Service Authorization Reference.

$sel:id:Organization', organization_id - The unique identifier (ID) of an organization.

The regex pattern for an organization ID string requires "o-" followed by from 10 to 32 lowercase letters or digits.

$sel:masterAccountId:Organization', organization_masterAccountId - The unique identifier (ID) of the management account of an organization.

The regex pattern for an account ID string requires exactly 12 digits.

$sel:availablePolicyTypes:Organization', organization_availablePolicyTypes - Do not use. This field is deprecated and doesn't provide complete information about the policies in your organization.

To determine the policies that are enabled and available for use in your organization, use the ListRoots operation instead.

organization_masterAccountEmail :: Lens' Organization (Maybe Text)

The email address that is associated with the AWS account that is designated as the management account for the organization.

organization_featureSet :: Lens' Organization (Maybe OrganizationFeatureSet)

Specifies the functionality that currently is available to the organization. If set to "ALL", then all features are enabled and policies can be applied to accounts in the organization. If set to "CONSOLIDATED_BILLING", then only consolidated billing functionality is available. For more information, see Enabling All Features in Your Organization in the AWS Organizations User Guide.

organization_masterAccountArn :: Lens' Organization (Maybe Text)

The Amazon Resource Name (ARN) of the account that is designated as the management account for the organization.

For more information about ARNs in Organizations, see ARN Formats Supported by Organizations in the AWS Service Authorization Reference.

organization_arn :: Lens' Organization (Maybe Text)

The Amazon Resource Name (ARN) of an organization.

For more information about ARNs in Organizations, see ARN Formats Supported by Organizations in the AWS Service Authorization Reference.

organization_id :: Lens' Organization (Maybe Text)

The unique identifier (ID) of an organization.

The regex pattern for an organization ID string requires "o-" followed by from 10 to 32 lowercase letters or digits.

organization_masterAccountId :: Lens' Organization (Maybe Text)

The unique identifier (ID) of the management account of an organization.

The regex pattern for an account ID string requires exactly 12 digits.

organization_availablePolicyTypes :: Lens' Organization (Maybe [PolicyTypeSummary])

Do not use. This field is deprecated and doesn't provide complete information about the policies in your organization.

To determine the policies that are enabled and available for use in your organization, use the ListRoots operation instead.