Network.AWS.Organizations.Types.Parent

Description

Documentation

data Parent

Contains information about either a root or an organizational unit (OU) that can contain OUs or accounts in an organization.

See: newParent smart constructor.

Constructors

Parent'

Fields

id :: Maybe Text

The unique identifier (ID) of the parent entity.

The regex pattern for a parent ID string requires one of the following:

Root - A string that begins with "r-" followed by from 4 to 32 lowercase letters or digits.

Organizational unit (OU) - A string that begins with "ou-" followed by from 4 to 32 lowercase letters or digits (the ID of the root that the OU is in). This string is followed by a second "-" dash and from 8 to 32 additional lowercase letters or digits.

type' :: Maybe ParentType

The type of the parent entity.

newParent :: Parent

Create a value of Parent with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:id:Parent', parent_id - The unique identifier (ID) of the parent entity.

The regex pattern for a parent ID string requires one of the following:

Root - A string that begins with "r-" followed by from 4 to 32 lowercase letters or digits.

Organizational unit (OU) - A string that begins with "ou-" followed by from 4 to 32 lowercase letters or digits (the ID of the root that the OU is in). This string is followed by a second "-" dash and from 8 to 32 additional lowercase letters or digits.

$sel:type':Parent', parent_type - The type of the parent entity.

parent_id :: Lens' Parent (Maybe Text)

The unique identifier (ID) of the parent entity.

The regex pattern for a parent ID string requires one of the following:

Root - A string that begins with "r-" followed by from 4 to 32 lowercase letters or digits.

Organizational unit (OU) - A string that begins with "ou-" followed by from 4 to 32 lowercase letters or digits (the ID of the root that the OU is in). This string is followed by a second "-" dash and from 8 to 32 additional lowercase letters or digits.

parent_type :: Lens' Parent (Maybe ParentType)

The type of the parent entity.