Network.AWS.Organizations.Types.PolicyTargetSummary

Description

Documentation

data PolicyTargetSummary

Contains information about a root, OU, or account that a policy is attached to.

See: newPolicyTargetSummary smart constructor.

Constructors

PolicyTargetSummary'

Fields

targetId :: Maybe Text

The unique identifier (ID) of the policy target.

The regex pattern for a target ID string requires one of the following:

Root - A string that begins with "r-" followed by from 4 to 32 lowercase letters or digits.

Account - A string that consists of exactly 12 digits.

Organizational unit (OU) - A string that begins with "ou-" followed by from 4 to 32 lowercase letters or digits (the ID of the root that the OU is in). This string is followed by a second "-" dash and from 8 to 32 additional lowercase letters or digits.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the policy target.

For more information about ARNs in Organizations, see ARN Formats Supported by Organizations in the AWS Service Authorization Reference.

name :: Maybe Text

The friendly name of the policy target.

The regex pattern that is used to validate this parameter is a string of any of the characters in the ASCII character range.

type' :: Maybe TargetType

The type of the policy target.

newPolicyTargetSummary :: PolicyTargetSummary

Create a value of PolicyTargetSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetId:PolicyTargetSummary', policyTargetSummary_targetId - The unique identifier (ID) of the policy target.

The regex pattern for a target ID string requires one of the following:

Root - A string that begins with "r-" followed by from 4 to 32 lowercase letters or digits.

Account - A string that consists of exactly 12 digits.

Organizational unit (OU) - A string that begins with "ou-" followed by from 4 to 32 lowercase letters or digits (the ID of the root that the OU is in). This string is followed by a second "-" dash and from 8 to 32 additional lowercase letters or digits.

$sel:arn:PolicyTargetSummary', policyTargetSummary_arn - The Amazon Resource Name (ARN) of the policy target.

For more information about ARNs in Organizations, see ARN Formats Supported by Organizations in the AWS Service Authorization Reference.

$sel:name:PolicyTargetSummary', policyTargetSummary_name - The friendly name of the policy target.

The regex pattern that is used to validate this parameter is a string of any of the characters in the ASCII character range.

$sel:type':PolicyTargetSummary', policyTargetSummary_type - The type of the policy target.

policyTargetSummary_targetId :: Lens' PolicyTargetSummary (Maybe Text)

The unique identifier (ID) of the policy target.

The regex pattern for a target ID string requires one of the following:

Root - A string that begins with "r-" followed by from 4 to 32 lowercase letters or digits.

Account - A string that consists of exactly 12 digits.

Organizational unit (OU) - A string that begins with "ou-" followed by from 4 to 32 lowercase letters or digits (the ID of the root that the OU is in). This string is followed by a second "-" dash and from 8 to 32 additional lowercase letters or digits.

policyTargetSummary_arn :: Lens' PolicyTargetSummary (Maybe Text)

The Amazon Resource Name (ARN) of the policy target.

For more information about ARNs in Organizations, see ARN Formats Supported by Organizations in the AWS Service Authorization Reference.

policyTargetSummary_name :: Lens' PolicyTargetSummary (Maybe Text)

The friendly name of the policy target.

The regex pattern that is used to validate this parameter is a string of any of the characters in the ASCII character range.

policyTargetSummary_type :: Lens' PolicyTargetSummary (Maybe TargetType)

The type of the policy target.