Network.AWS.Organizations.UpdatePolicy

Description

Updates an existing policy with a new name, description, or content. If you don't supply any parameter, that value remains unchanged. You can't change a policy's type.

This operation can be called only from the organization's management account.

Creating a Request

data UpdatePolicy

See: newUpdatePolicy smart constructor.

Constructors

UpdatePolicy'

Fields

content :: Maybe Text

If provided, the new content for the policy. The text must be correctly formatted JSON that complies with the syntax for the policy's type. For more information, see Service Control Policy Syntax in the AWS Organizations User Guide.

name :: Maybe Text

If provided, the new name for the policy.

The regex pattern that is used to validate this parameter is a string of any of the characters in the ASCII character range.

description :: Maybe Text

If provided, the new description for the policy.

policyId :: Text

The unique identifier (ID) of the policy that you want to update.

The regex pattern for a policy ID string requires "p-" followed by from 8 to 128 lowercase or uppercase letters, digits, or the underscore character (_).

newUpdatePolicy

Arguments

:: Text
$sel:policyId:UpdatePolicy'

-> UpdatePolicy

Create a value of UpdatePolicy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:content:UpdatePolicy', updatePolicy_content - If provided, the new content for the policy. The text must be correctly formatted JSON that complies with the syntax for the policy's type. For more information, see Service Control Policy Syntax in the AWS Organizations User Guide.

$sel:name:UpdatePolicy', updatePolicy_name - If provided, the new name for the policy.

The regex pattern that is used to validate this parameter is a string of any of the characters in the ASCII character range.

$sel:description:UpdatePolicy', updatePolicy_description - If provided, the new description for the policy.

$sel:policyId:UpdatePolicy', updatePolicy_policyId - The unique identifier (ID) of the policy that you want to update.

The regex pattern for a policy ID string requires "p-" followed by from 8 to 128 lowercase or uppercase letters, digits, or the underscore character (_).

Request Lenses

updatePolicy_content :: Lens' UpdatePolicy (Maybe Text)

If provided, the new content for the policy. The text must be correctly formatted JSON that complies with the syntax for the policy's type. For more information, see Service Control Policy Syntax in the AWS Organizations User Guide.

updatePolicy_name :: Lens' UpdatePolicy (Maybe Text)

If provided, the new name for the policy.

The regex pattern that is used to validate this parameter is a string of any of the characters in the ASCII character range.

updatePolicy_description :: Lens' UpdatePolicy (Maybe Text)

If provided, the new description for the policy.

updatePolicy_policyId :: Lens' UpdatePolicy Text

The unique identifier (ID) of the policy that you want to update.

The regex pattern for a policy ID string requires "p-" followed by from 8 to 128 lowercase or uppercase letters, digits, or the underscore character (_).

Destructuring the Response

data UpdatePolicyResponse

See: newUpdatePolicyResponse smart constructor.

Constructors

UpdatePolicyResponse'

Fields

policy :: Maybe Policy

A structure that contains details about the updated policy, showing the requested changes.

httpStatus :: Int

The response's http status code.

newUpdatePolicyResponse

Arguments

:: Int
$sel:httpStatus:UpdatePolicyResponse'

-> UpdatePolicyResponse

Create a value of UpdatePolicyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:policy:UpdatePolicyResponse', updatePolicyResponse_policy - A structure that contains details about the updated policy, showing the requested changes.

$sel:httpStatus:UpdatePolicyResponse', updatePolicyResponse_httpStatus - The response's http status code.

Response Lenses

updatePolicyResponse_policy :: Lens' UpdatePolicyResponse (Maybe Policy)

A structure that contains details about the updated policy, showing the requested changes.

updatePolicyResponse_httpStatus :: Lens' UpdatePolicyResponse Int

The response's http status code.