Network.AWS.Outposts.CreateOrder

Description

Creates an order for an Outpost.

Creating a Request

data CreateOrder

See: newCreateOrder smart constructor.

Constructors

CreateOrder'

Fields

paymentTerm :: Maybe PaymentTerm

The payment terms for the order.

outpostIdentifier :: Text

The ID or the Amazon Resource Name (ARN) of the Outpost.

lineItems :: NonEmpty LineItemRequest

The line items that make up the order.

paymentOption :: PaymentOption

The payment option for the order.

newCreateOrder

Arguments

:: Text
$sel:outpostIdentifier:CreateOrder'

-> NonEmpty LineItemRequest
$sel:lineItems:CreateOrder'

-> PaymentOption
$sel:paymentOption:CreateOrder'

-> CreateOrder

Create a value of CreateOrder with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:paymentTerm:CreateOrder', createOrder_paymentTerm - The payment terms for the order.

$sel:outpostIdentifier:CreateOrder', createOrder_outpostIdentifier - The ID or the Amazon Resource Name (ARN) of the Outpost.

$sel:lineItems:CreateOrder', createOrder_lineItems - The line items that make up the order.

$sel:paymentOption:CreateOrder', createOrder_paymentOption - The payment option for the order.

Request Lenses

createOrder_paymentTerm :: Lens' CreateOrder (Maybe PaymentTerm)

The payment terms for the order.

createOrder_outpostIdentifier :: Lens' CreateOrder Text

The ID or the Amazon Resource Name (ARN) of the Outpost.

createOrder_lineItems :: Lens' CreateOrder (NonEmpty LineItemRequest)

The line items that make up the order.

createOrder_paymentOption :: Lens' CreateOrder PaymentOption

The payment option for the order.

Destructuring the Response

data CreateOrderResponse

See: newCreateOrderResponse smart constructor.

Constructors

CreateOrderResponse'

Fields

order :: Maybe Order

Information about this order.

httpStatus :: Int

The response's http status code.

newCreateOrderResponse

Arguments

:: Int
$sel:httpStatus:CreateOrderResponse'

-> CreateOrderResponse

Create a value of CreateOrderResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:order:CreateOrderResponse', createOrderResponse_order - Information about this order.

$sel:httpStatus:CreateOrderResponse', createOrderResponse_httpStatus - The response's http status code.

Response Lenses

createOrderResponse_order :: Lens' CreateOrderResponse (Maybe Order)

Information about this order.

createOrderResponse_httpStatus :: Lens' CreateOrderResponse Int

The response's http status code.