Network.AWS.Outposts.GetOutpostInstanceTypes

Description

Lists the instance types for the specified Outpost.

Creating a Request

data GetOutpostInstanceTypes

See: newGetOutpostInstanceTypes smart constructor.

Constructors

GetOutpostInstanceTypes'

Fields

nextToken :: Maybe Text

maxResults :: Maybe Natural

outpostId :: Text

The ID of the Outpost.

newGetOutpostInstanceTypes

Arguments

:: Text
$sel:outpostId:GetOutpostInstanceTypes'

-> GetOutpostInstanceTypes

Create a value of GetOutpostInstanceTypes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetOutpostInstanceTypes', getOutpostInstanceTypes_nextToken - Undocumented member.

$sel:maxResults:GetOutpostInstanceTypes', getOutpostInstanceTypes_maxResults - Undocumented member.

$sel:outpostId:GetOutpostInstanceTypes', getOutpostInstanceTypes_outpostId - The ID of the Outpost.

Request Lenses

getOutpostInstanceTypes_nextToken :: Lens' GetOutpostInstanceTypes (Maybe Text)

Undocumented member.

getOutpostInstanceTypes_maxResults :: Lens' GetOutpostInstanceTypes (Maybe Natural)

Undocumented member.

getOutpostInstanceTypes_outpostId :: Lens' GetOutpostInstanceTypes Text

The ID of the Outpost.

Destructuring the Response

data GetOutpostInstanceTypesResponse

See: newGetOutpostInstanceTypesResponse smart constructor.

Constructors

GetOutpostInstanceTypesResponse'

Fields

instanceTypes :: Maybe [InstanceTypeItem]

outpostArn :: Maybe Text

nextToken :: Maybe Text

outpostId :: Maybe Text

The ID of the Outpost.

httpStatus :: Int

The response's http status code.

newGetOutpostInstanceTypesResponse

Arguments

:: Int
$sel:httpStatus:GetOutpostInstanceTypesResponse'

-> GetOutpostInstanceTypesResponse

Create a value of GetOutpostInstanceTypesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceTypes:GetOutpostInstanceTypesResponse', getOutpostInstanceTypesResponse_instanceTypes - Undocumented member.

$sel:outpostArn:GetOutpostInstanceTypesResponse', getOutpostInstanceTypesResponse_outpostArn - Undocumented member.

$sel:nextToken:GetOutpostInstanceTypes', getOutpostInstanceTypesResponse_nextToken - Undocumented member.

$sel:outpostId:GetOutpostInstanceTypes', getOutpostInstanceTypesResponse_outpostId - The ID of the Outpost.

$sel:httpStatus:GetOutpostInstanceTypesResponse', getOutpostInstanceTypesResponse_httpStatus - The response's http status code.

Response Lenses

getOutpostInstanceTypesResponse_instanceTypes :: Lens' GetOutpostInstanceTypesResponse (Maybe [InstanceTypeItem])

Undocumented member.

getOutpostInstanceTypesResponse_outpostArn :: Lens' GetOutpostInstanceTypesResponse (Maybe Text)

Undocumented member.

getOutpostInstanceTypesResponse_nextToken :: Lens' GetOutpostInstanceTypesResponse (Maybe Text)

Undocumented member.

getOutpostInstanceTypesResponse_outpostId :: Lens' GetOutpostInstanceTypesResponse (Maybe Text)

The ID of the Outpost.

getOutpostInstanceTypesResponse_httpStatus :: Lens' GetOutpostInstanceTypesResponse Int

The response's http status code.