Network.AWS.Outposts.Lens

Description

Operations

DeleteOutpost

deleteOutpost_outpostId :: Lens' DeleteOutpost Text

The ID of the Outpost.

deleteOutpostResponse_httpStatus :: Lens' DeleteOutpostResponse Int

The response's http status code.

DeleteSite

deleteSite_siteId :: Lens' DeleteSite Text

Undocumented member.

deleteSiteResponse_httpStatus :: Lens' DeleteSiteResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) of the resource.

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The resource tags.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

ListOutposts

listOutposts_availabilityZoneFilter :: Lens' ListOutposts (Maybe (NonEmpty Text))

A filter for the Availibility Zone (us-east-1a) of the Outpost.

Filter values are case sensitive. If you specify multiple values for a filter, the values are joined with an OR, and the request returns all results that match any of the specified values.

listOutposts_lifeCycleStatusFilter :: Lens' ListOutposts (Maybe (NonEmpty Text))

A filter for the lifecycle status of the Outpost.

Filter values are case sensitive. If you specify multiple values for a filter, the values are joined with an OR, and the request returns all results that match any of the specified values.

listOutposts_nextToken :: Lens' ListOutposts (Maybe Text)

Undocumented member.

listOutposts_availabilityZoneIdFilter :: Lens' ListOutposts (Maybe (NonEmpty Text))

A filter for the AZ IDs (use1-az1) of the Outpost.

Filter values are case sensitive. If you specify multiple values for a filter, the values are joined with an OR, and the request returns all results that match any of the specified values.

listOutposts_maxResults :: Lens' ListOutposts (Maybe Natural)

Undocumented member.

listOutpostsResponse_nextToken :: Lens' ListOutpostsResponse (Maybe Text)

Undocumented member.

listOutpostsResponse_outposts :: Lens' ListOutpostsResponse (Maybe [Outpost])

Undocumented member.

listOutpostsResponse_httpStatus :: Lens' ListOutpostsResponse Int

The response's http status code.

ListSites

listSites_nextToken :: Lens' ListSites (Maybe Text)

Undocumented member.

listSites_maxResults :: Lens' ListSites (Maybe Natural)

Undocumented member.

listSitesResponse_nextToken :: Lens' ListSitesResponse (Maybe Text)

Undocumented member.

listSitesResponse_sites :: Lens' ListSitesResponse (Maybe [Site])

Undocumented member.

listSitesResponse_httpStatus :: Lens' ListSitesResponse Int

The response's http status code.

CreateOrder

createOrder_paymentTerm :: Lens' CreateOrder (Maybe PaymentTerm)

The payment terms for the order.

createOrder_outpostIdentifier :: Lens' CreateOrder Text

The ID or the Amazon Resource Name (ARN) of the Outpost.

createOrder_lineItems :: Lens' CreateOrder (NonEmpty LineItemRequest)

The line items that make up the order.

createOrder_paymentOption :: Lens' CreateOrder PaymentOption

The payment option for the order.

createOrderResponse_order :: Lens' CreateOrderResponse (Maybe Order)

Information about this order.

createOrderResponse_httpStatus :: Lens' CreateOrderResponse Int

The response's http status code.

GetOutpostInstanceTypes

getOutpostInstanceTypes_nextToken :: Lens' GetOutpostInstanceTypes (Maybe Text)

Undocumented member.

getOutpostInstanceTypes_maxResults :: Lens' GetOutpostInstanceTypes (Maybe Natural)

Undocumented member.

getOutpostInstanceTypes_outpostId :: Lens' GetOutpostInstanceTypes Text

The ID of the Outpost.

getOutpostInstanceTypesResponse_instanceTypes :: Lens' GetOutpostInstanceTypesResponse (Maybe [InstanceTypeItem])

Undocumented member.

getOutpostInstanceTypesResponse_outpostArn :: Lens' GetOutpostInstanceTypesResponse (Maybe Text)

Undocumented member.

getOutpostInstanceTypesResponse_nextToken :: Lens' GetOutpostInstanceTypesResponse (Maybe Text)

Undocumented member.

getOutpostInstanceTypesResponse_outpostId :: Lens' GetOutpostInstanceTypesResponse (Maybe Text)

The ID of the Outpost.

getOutpostInstanceTypesResponse_httpStatus :: Lens' GetOutpostInstanceTypesResponse Int

The response's http status code.

CreateOutpost

createOutpost_availabilityZoneId :: Lens' CreateOutpost (Maybe Text)

Undocumented member.

createOutpost_availabilityZone :: Lens' CreateOutpost (Maybe Text)

Undocumented member.

createOutpost_description :: Lens' CreateOutpost (Maybe Text)

Undocumented member.

createOutpost_tags :: Lens' CreateOutpost (Maybe (HashMap Text Text))

The tags to apply to the Outpost.

createOutpost_name :: Lens' CreateOutpost Text

Undocumented member.

createOutpost_siteId :: Lens' CreateOutpost Text

Undocumented member.

createOutpostResponse_outpost :: Lens' CreateOutpostResponse (Maybe Outpost)

Undocumented member.

createOutpostResponse_httpStatus :: Lens' CreateOutpostResponse Int

The response's http status code.

GetOutpost

getOutpost_outpostId :: Lens' GetOutpost Text

The ID of the Outpost.

getOutpostResponse_outpost :: Lens' GetOutpostResponse (Maybe Outpost)

Undocumented member.

getOutpostResponse_httpStatus :: Lens' GetOutpostResponse Int

The response's http status code.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) of the resource.

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The tags to add to the resource.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the resource.

untagResource_tagKeys :: Lens' UntagResource (NonEmpty Text)

The tag keys.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

Types

InstanceTypeItem

instanceTypeItem_instanceType :: Lens' InstanceTypeItem (Maybe Text)

Undocumented member.

LineItem

lineItem_status :: Lens' LineItem (Maybe Text)

The status of the line item.

lineItem_quantity :: Lens' LineItem (Maybe Natural)

The quantity of the line item.

lineItem_catalogItemId :: Lens' LineItem (Maybe Text)

The ID of the catalog item.

lineItem_lineItemId :: Lens' LineItem (Maybe Text)

The ID of the line item.

LineItemRequest

lineItemRequest_quantity :: Lens' LineItemRequest (Maybe Natural)

The quantity of a line item request.

lineItemRequest_catalogItemId :: Lens' LineItemRequest (Maybe Text)

The ID of the catalog item.

Order

order_status :: Lens' Order (Maybe OrderStatus)

The status of the order

order_orderSubmissionDate :: Lens' Order (Maybe UTCTime)

The submission date for the order.

order_lineItems :: Lens' Order (Maybe [LineItem])

The line items for the order

order_orderFulfilledDate :: Lens' Order (Maybe UTCTime)

The fulfillment date of the order.

order_orderId :: Lens' Order (Maybe Text)

The ID of the order.

order_outpostId :: Lens' Order (Maybe Text)

The ID of the Outpost.

order_paymentOption :: Lens' Order (Maybe PaymentOption)

The payment option for the order.

Outpost

outpost_availabilityZoneId :: Lens' Outpost (Maybe Text)

Undocumented member.

outpost_outpostArn :: Lens' Outpost (Maybe Text)

Undocumented member.

outpost_ownerId :: Lens' Outpost (Maybe Text)

Undocumented member.

outpost_availabilityZone :: Lens' Outpost (Maybe Text)

Undocumented member.

outpost_name :: Lens' Outpost (Maybe Text)

Undocumented member.

outpost_lifeCycleStatus :: Lens' Outpost (Maybe Text)

Undocumented member.

outpost_outpostId :: Lens' Outpost (Maybe Text)

The ID of the Outpost.

outpost_siteId :: Lens' Outpost (Maybe Text)

Undocumented member.

outpost_siteArn :: Lens' Outpost (Maybe Text)

Undocumented member.

outpost_description :: Lens' Outpost (Maybe Text)

Undocumented member.

outpost_tags :: Lens' Outpost (Maybe (HashMap Text Text))

The Outpost tags.

Site

site_accountId :: Lens' Site (Maybe Text)

Undocumented member.

site_name :: Lens' Site (Maybe Text)

Undocumented member.

site_siteId :: Lens' Site (Maybe Text)

Undocumented member.

site_siteArn :: Lens' Site (Maybe Text)

Undocumented member.

site_description :: Lens' Site (Maybe Text)

Undocumented member.

site_tags :: Lens' Site (Maybe (HashMap Text Text))

The site tags.