Network.AWS.Outposts.Types

Description

Service Configuration

defaultService :: Service

API version 2019-12-03 of the Amazon Outposts SDK configuration.

Errors

_ValidationException :: AsError a => Getting (First ServiceError) a ServiceError

A parameter is not valid.

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have permission to perform this operation.

_ConflictException :: AsError a => Getting (First ServiceError) a ServiceError

Updating or deleting this resource can cause an inconsistent state.

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified request is not valid.

_ServiceQuotaExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You have exceeded a service quota.

_InternalServerException :: AsError a => Getting (First ServiceError) a ServiceError

An internal error has occurred.

OrderStatus

newtype OrderStatus

Constructors

OrderStatus'

Fields

fromOrderStatus :: Text

Bundled Patterns

pattern OrderStatus_CANCELLED :: OrderStatus

pattern OrderStatus_FULFILLED :: OrderStatus

pattern OrderStatus_INSTALLING :: OrderStatus

pattern OrderStatus_PENDING :: OrderStatus

pattern OrderStatus_PROCESSING :: OrderStatus

pattern OrderStatus_RECEIVED :: OrderStatus

PaymentOption

newtype PaymentOption

Constructors

PaymentOption'

Fields

fromPaymentOption :: Text

Bundled Patterns

pattern PaymentOption_ALL_UPFRONT :: PaymentOption

pattern PaymentOption_NO_UPFRONT :: PaymentOption

pattern PaymentOption_PARTIAL_UPFRONT :: PaymentOption

PaymentTerm

newtype PaymentTerm

Constructors

PaymentTerm'

Fields

fromPaymentTerm :: Text

Bundled Patterns

pattern PaymentTerm_THREE_YEARS :: PaymentTerm

InstanceTypeItem

data InstanceTypeItem

Information about an instance type.

See: newInstanceTypeItem smart constructor.

Constructors

InstanceTypeItem'

Fields

instanceType :: Maybe Text

newInstanceTypeItem :: InstanceTypeItem

Create a value of InstanceTypeItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:instanceType:InstanceTypeItem', instanceTypeItem_instanceType - Undocumented member.

instanceTypeItem_instanceType :: Lens' InstanceTypeItem (Maybe Text)

Undocumented member.

LineItem

data LineItem

Information about a line item.

See: newLineItem smart constructor.

Constructors

LineItem'

Fields

status :: Maybe Text

The status of the line item.

quantity :: Maybe Natural

The quantity of the line item.

catalogItemId :: Maybe Text

The ID of the catalog item.

lineItemId :: Maybe Text

The ID of the line item.

newLineItem :: LineItem

Create a value of LineItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:LineItem', lineItem_status - The status of the line item.

$sel:quantity:LineItem', lineItem_quantity - The quantity of the line item.

$sel:catalogItemId:LineItem', lineItem_catalogItemId - The ID of the catalog item.

$sel:lineItemId:LineItem', lineItem_lineItemId - The ID of the line item.

lineItem_status :: Lens' LineItem (Maybe Text)

The status of the line item.

lineItem_quantity :: Lens' LineItem (Maybe Natural)

The quantity of the line item.

lineItem_catalogItemId :: Lens' LineItem (Maybe Text)

The ID of the catalog item.

lineItem_lineItemId :: Lens' LineItem (Maybe Text)

The ID of the line item.

LineItemRequest

data LineItemRequest

Information about a line item request.

See: newLineItemRequest smart constructor.

Constructors

LineItemRequest'

Fields

quantity :: Maybe Natural

The quantity of a line item request.

catalogItemId :: Maybe Text

The ID of the catalog item.

newLineItemRequest :: LineItemRequest

Create a value of LineItemRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:quantity:LineItemRequest', lineItemRequest_quantity - The quantity of a line item request.

$sel:catalogItemId:LineItemRequest', lineItemRequest_catalogItemId - The ID of the catalog item.

lineItemRequest_quantity :: Lens' LineItemRequest (Maybe Natural)

The quantity of a line item request.

lineItemRequest_catalogItemId :: Lens' LineItemRequest (Maybe Text)

The ID of the catalog item.

Order

data Order

Information about an order.

See: newOrder smart constructor.

Constructors

Order'

Fields

status :: Maybe OrderStatus

The status of the order

orderSubmissionDate :: Maybe POSIX

The submission date for the order.

lineItems :: Maybe [LineItem]

The line items for the order

orderFulfilledDate :: Maybe POSIX

The fulfillment date of the order.

orderId :: Maybe Text

The ID of the order.

outpostId :: Maybe Text

The ID of the Outpost.

paymentOption :: Maybe PaymentOption

The payment option for the order.

newOrder :: Order

Create a value of Order with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Order', order_status - The status of the order

$sel:orderSubmissionDate:Order', order_orderSubmissionDate - The submission date for the order.

$sel:lineItems:Order', order_lineItems - The line items for the order

$sel:orderFulfilledDate:Order', order_orderFulfilledDate - The fulfillment date of the order.

$sel:orderId:Order', order_orderId - The ID of the order.

$sel:outpostId:Order', order_outpostId - The ID of the Outpost.

$sel:paymentOption:Order', order_paymentOption - The payment option for the order.

order_status :: Lens' Order (Maybe OrderStatus)

The status of the order

order_orderSubmissionDate :: Lens' Order (Maybe UTCTime)

The submission date for the order.

order_lineItems :: Lens' Order (Maybe [LineItem])

The line items for the order

order_orderFulfilledDate :: Lens' Order (Maybe UTCTime)

The fulfillment date of the order.

order_orderId :: Lens' Order (Maybe Text)

The ID of the order.

order_outpostId :: Lens' Order (Maybe Text)

The ID of the Outpost.

order_paymentOption :: Lens' Order (Maybe PaymentOption)

The payment option for the order.

Outpost

data Outpost

Information about an Outpost.

See: newOutpost smart constructor.

Constructors

Outpost'

Fields

availabilityZoneId :: Maybe Text

outpostArn :: Maybe Text

ownerId :: Maybe Text

availabilityZone :: Maybe Text

name :: Maybe Text

lifeCycleStatus :: Maybe Text

outpostId :: Maybe Text

The ID of the Outpost.

siteId :: Maybe Text

siteArn :: Maybe Text

description :: Maybe Text

tags :: Maybe (HashMap Text Text)

The Outpost tags.

newOutpost :: Outpost

Create a value of Outpost with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:availabilityZoneId:Outpost', outpost_availabilityZoneId - Undocumented member.

$sel:outpostArn:Outpost', outpost_outpostArn - Undocumented member.

$sel:ownerId:Outpost', outpost_ownerId - Undocumented member.

$sel:availabilityZone:Outpost', outpost_availabilityZone - Undocumented member.

$sel:name:Outpost', outpost_name - Undocumented member.

$sel:lifeCycleStatus:Outpost', outpost_lifeCycleStatus - Undocumented member.

$sel:outpostId:Outpost', outpost_outpostId - The ID of the Outpost.

$sel:siteId:Outpost', outpost_siteId - Undocumented member.

$sel:siteArn:Outpost', outpost_siteArn - Undocumented member.

$sel:description:Outpost', outpost_description - Undocumented member.

$sel:tags:Outpost', outpost_tags - The Outpost tags.

outpost_availabilityZoneId :: Lens' Outpost (Maybe Text)

Undocumented member.

outpost_outpostArn :: Lens' Outpost (Maybe Text)

Undocumented member.

outpost_ownerId :: Lens' Outpost (Maybe Text)

Undocumented member.

outpost_availabilityZone :: Lens' Outpost (Maybe Text)

Undocumented member.

outpost_name :: Lens' Outpost (Maybe Text)

Undocumented member.

outpost_lifeCycleStatus :: Lens' Outpost (Maybe Text)

Undocumented member.

outpost_outpostId :: Lens' Outpost (Maybe Text)

The ID of the Outpost.

outpost_siteId :: Lens' Outpost (Maybe Text)

Undocumented member.

outpost_siteArn :: Lens' Outpost (Maybe Text)

Undocumented member.

outpost_description :: Lens' Outpost (Maybe Text)

Undocumented member.

outpost_tags :: Lens' Outpost (Maybe (HashMap Text Text))

The Outpost tags.

Site

data Site

Information about a site.

See: newSite smart constructor.

Constructors

Site'

Fields

accountId :: Maybe Text

name :: Maybe Text

siteId :: Maybe Text

siteArn :: Maybe Text

description :: Maybe Text

tags :: Maybe (HashMap Text Text)

The site tags.

newSite :: Site

Create a value of Site with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountId:Site', site_accountId - Undocumented member.

$sel:name:Site', site_name - Undocumented member.

$sel:siteId:Site', site_siteId - Undocumented member.

$sel:siteArn:Site', site_siteArn - Undocumented member.

$sel:description:Site', site_description - Undocumented member.

$sel:tags:Site', site_tags - The site tags.

site_accountId :: Lens' Site (Maybe Text)

Undocumented member.

site_name :: Lens' Site (Maybe Text)

Undocumented member.

site_siteId :: Lens' Site (Maybe Text)

Undocumented member.

site_siteArn :: Lens' Site (Maybe Text)

Undocumented member.

site_description :: Lens' Site (Maybe Text)

Undocumented member.

site_tags :: Lens' Site (Maybe (HashMap Text Text))

The site tags.