Network.AWS.Panorama.DescribeApplicationInstanceDetails

Description

Returns information about an application instance's configuration manifest.

Creating a Request

data DescribeApplicationInstanceDetails

See: newDescribeApplicationInstanceDetails smart constructor.

Constructors

DescribeApplicationInstanceDetails'

Fields

applicationInstanceId :: Text

The application instance's ID.

newDescribeApplicationInstanceDetails

Arguments

:: Text
$sel:applicationInstanceId:DescribeApplicationInstanceDetails'

-> DescribeApplicationInstanceDetails

Create a value of DescribeApplicationInstanceDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationInstanceId:DescribeApplicationInstanceDetails', describeApplicationInstanceDetails_applicationInstanceId - The application instance's ID.

Request Lenses

describeApplicationInstanceDetails_applicationInstanceId :: Lens' DescribeApplicationInstanceDetails Text

The application instance's ID.

Destructuring the Response

data DescribeApplicationInstanceDetailsResponse

See: newDescribeApplicationInstanceDetailsResponse smart constructor.

Constructors

DescribeApplicationInstanceDetailsResponse'

Fields

createdTime :: Maybe POSIX

When the application instance was created.

defaultRuntimeContextDevice :: Maybe Text

The application instance's default runtime context device.

manifestOverridesPayload :: Maybe ManifestOverridesPayload

Parameter overrides for the configuration manifest.

name :: Maybe Text

The application instance's name.

applicationInstanceId :: Maybe Text

The application instance's ID.

description :: Maybe Text

The application instance's description.

manifestPayload :: Maybe ManifestPayload

The application instance's configuration manifest.

applicationInstanceIdToReplace :: Maybe Text

The ID of the application instance that this instance replaced.

httpStatus :: Int

The response's http status code.

newDescribeApplicationInstanceDetailsResponse

Arguments

:: Int
$sel:httpStatus:DescribeApplicationInstanceDetailsResponse'

-> DescribeApplicationInstanceDetailsResponse

Create a value of DescribeApplicationInstanceDetailsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:createdTime:DescribeApplicationInstanceDetailsResponse', describeApplicationInstanceDetailsResponse_createdTime - When the application instance was created.

$sel:defaultRuntimeContextDevice:DescribeApplicationInstanceDetailsResponse', describeApplicationInstanceDetailsResponse_defaultRuntimeContextDevice - The application instance's default runtime context device.

$sel:manifestOverridesPayload:DescribeApplicationInstanceDetailsResponse', describeApplicationInstanceDetailsResponse_manifestOverridesPayload - Parameter overrides for the configuration manifest.

$sel:name:DescribeApplicationInstanceDetailsResponse', describeApplicationInstanceDetailsResponse_name - The application instance's name.

$sel:applicationInstanceId:DescribeApplicationInstanceDetails', describeApplicationInstanceDetailsResponse_applicationInstanceId - The application instance's ID.

$sel:description:DescribeApplicationInstanceDetailsResponse', describeApplicationInstanceDetailsResponse_description - The application instance's description.

$sel:manifestPayload:DescribeApplicationInstanceDetailsResponse', describeApplicationInstanceDetailsResponse_manifestPayload - The application instance's configuration manifest.

$sel:applicationInstanceIdToReplace:DescribeApplicationInstanceDetailsResponse', describeApplicationInstanceDetailsResponse_applicationInstanceIdToReplace - The ID of the application instance that this instance replaced.

$sel:httpStatus:DescribeApplicationInstanceDetailsResponse', describeApplicationInstanceDetailsResponse_httpStatus - The response's http status code.

Response Lenses

describeApplicationInstanceDetailsResponse_createdTime :: Lens' DescribeApplicationInstanceDetailsResponse (Maybe UTCTime)

When the application instance was created.

describeApplicationInstanceDetailsResponse_defaultRuntimeContextDevice :: Lens' DescribeApplicationInstanceDetailsResponse (Maybe Text)

The application instance's default runtime context device.

describeApplicationInstanceDetailsResponse_manifestOverridesPayload :: Lens' DescribeApplicationInstanceDetailsResponse (Maybe ManifestOverridesPayload)

Parameter overrides for the configuration manifest.

describeApplicationInstanceDetailsResponse_name :: Lens' DescribeApplicationInstanceDetailsResponse (Maybe Text)

The application instance's name.

describeApplicationInstanceDetailsResponse_applicationInstanceId :: Lens' DescribeApplicationInstanceDetailsResponse (Maybe Text)

The application instance's ID.

describeApplicationInstanceDetailsResponse_description :: Lens' DescribeApplicationInstanceDetailsResponse (Maybe Text)

The application instance's description.

describeApplicationInstanceDetailsResponse_manifestPayload :: Lens' DescribeApplicationInstanceDetailsResponse (Maybe ManifestPayload)

The application instance's configuration manifest.

describeApplicationInstanceDetailsResponse_applicationInstanceIdToReplace :: Lens' DescribeApplicationInstanceDetailsResponse (Maybe Text)

The ID of the application instance that this instance replaced.

describeApplicationInstanceDetailsResponse_httpStatus :: Lens' DescribeApplicationInstanceDetailsResponse Int

The response's http status code.