Network.AWS.Panorama.DescribeDeviceJob

Description

Returns information about a device job.

Creating a Request

data DescribeDeviceJob

See: newDescribeDeviceJob smart constructor.

Constructors

DescribeDeviceJob'

Fields

jobId :: Text

The job's ID.

newDescribeDeviceJob

Arguments

:: Text
$sel:jobId:DescribeDeviceJob'

-> DescribeDeviceJob

Create a value of DescribeDeviceJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:DescribeDeviceJob', describeDeviceJob_jobId - The job's ID.

Request Lenses

describeDeviceJob_jobId :: Lens' DescribeDeviceJob Text

The job's ID.

Destructuring the Response

data DescribeDeviceJobResponse

See: newDescribeDeviceJobResponse smart constructor.

Constructors

DescribeDeviceJobResponse'

Fields

status :: Maybe UpdateProgress

The job's status.

jobId :: Maybe Text

The job's ID.

createdTime :: Maybe POSIX

When the job was created.

deviceArn :: Maybe Text

The device's ARN.

imageVersion :: Maybe Text

For an OTA job, the target version of the device software.

deviceName :: Maybe Text

The device's name.

deviceId :: Maybe Text

The device's ID.

deviceType :: Maybe DeviceType

The device's type.

httpStatus :: Int

The response's http status code.

newDescribeDeviceJobResponse

Arguments

:: Int
$sel:httpStatus:DescribeDeviceJobResponse'

-> DescribeDeviceJobResponse

Create a value of DescribeDeviceJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeDeviceJobResponse', describeDeviceJobResponse_status - The job's status.

$sel:jobId:DescribeDeviceJob', describeDeviceJobResponse_jobId - The job's ID.

$sel:createdTime:DescribeDeviceJobResponse', describeDeviceJobResponse_createdTime - When the job was created.

$sel:deviceArn:DescribeDeviceJobResponse', describeDeviceJobResponse_deviceArn - The device's ARN.

$sel:imageVersion:DescribeDeviceJobResponse', describeDeviceJobResponse_imageVersion - For an OTA job, the target version of the device software.

$sel:deviceName:DescribeDeviceJobResponse', describeDeviceJobResponse_deviceName - The device's name.

$sel:deviceId:DescribeDeviceJobResponse', describeDeviceJobResponse_deviceId - The device's ID.

$sel:deviceType:DescribeDeviceJobResponse', describeDeviceJobResponse_deviceType - The device's type.

$sel:httpStatus:DescribeDeviceJobResponse', describeDeviceJobResponse_httpStatus - The response's http status code.

Response Lenses

describeDeviceJobResponse_status :: Lens' DescribeDeviceJobResponse (Maybe UpdateProgress)

The job's status.

describeDeviceJobResponse_jobId :: Lens' DescribeDeviceJobResponse (Maybe Text)

The job's ID.

describeDeviceJobResponse_createdTime :: Lens' DescribeDeviceJobResponse (Maybe UTCTime)

When the job was created.

describeDeviceJobResponse_deviceArn :: Lens' DescribeDeviceJobResponse (Maybe Text)

The device's ARN.

describeDeviceJobResponse_imageVersion :: Lens' DescribeDeviceJobResponse (Maybe Text)

For an OTA job, the target version of the device software.

describeDeviceJobResponse_deviceName :: Lens' DescribeDeviceJobResponse (Maybe Text)

The device's name.

describeDeviceJobResponse_deviceId :: Lens' DescribeDeviceJobResponse (Maybe Text)

The device's ID.

describeDeviceJobResponse_deviceType :: Lens' DescribeDeviceJobResponse (Maybe DeviceType)

The device's type.

describeDeviceJobResponse_httpStatus :: Lens' DescribeDeviceJobResponse Int

The response's http status code.