Network.AWS.Panorama.DescribeNode

Description

Returns information about a node.

Creating a Request

data DescribeNode

See: newDescribeNode smart constructor.

Constructors

DescribeNode'

Fields

ownerAccount :: Maybe Text

The account ID of the node's owner.

nodeId :: Text

The node's ID.

newDescribeNode

Arguments

:: Text
$sel:nodeId:DescribeNode'

-> DescribeNode

Create a value of DescribeNode with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ownerAccount:DescribeNode', describeNode_ownerAccount - The account ID of the node's owner.

$sel:nodeId:DescribeNode', describeNode_nodeId - The node's ID.

Request Lenses

describeNode_ownerAccount :: Lens' DescribeNode (Maybe Text)

The account ID of the node's owner.

describeNode_nodeId :: Lens' DescribeNode Text

The node's ID.

Destructuring the Response

data DescribeNodeResponse

See: newDescribeNodeResponse smart constructor.

Constructors

DescribeNodeResponse'

Fields

assetName :: Maybe Text

The node's asset name.

packageArn :: Maybe Text

The node's ARN.

httpStatus :: Int

The response's http status code.

nodeId :: Text

The node's ID.

name :: Text

The node's name.

category :: NodeCategory

The node's category.

ownerAccount :: Text

The account ID of the node's owner.

packageName :: Text

The node's package name.

packageId :: Text

The node's package ID.

packageVersion :: Text

The node's package version.

patchVersion :: Text

The node's patch version.

nodeInterface :: NodeInterface

The node's interface.

description :: Text

The node's description.

createdTime :: POSIX

When the node was created.

lastUpdatedTime :: POSIX

When the node was updated.

newDescribeNodeResponse

Arguments

:: Int
$sel:httpStatus:DescribeNodeResponse'

-> Text
$sel:nodeId:DescribeNode'

-> Text
$sel:name:DescribeNodeResponse'

-> NodeCategory
$sel:category:DescribeNodeResponse'

-> Text
$sel:ownerAccount:DescribeNode'

-> Text
$sel:packageName:DescribeNodeResponse'

-> Text
$sel:packageId:DescribeNodeResponse'

-> Text
$sel:packageVersion:DescribeNodeResponse'

-> Text
$sel:patchVersion:DescribeNodeResponse'

-> NodeInterface
$sel:nodeInterface:DescribeNodeResponse'

-> Text
$sel:description:DescribeNodeResponse'

-> UTCTime
$sel:createdTime:DescribeNodeResponse'

-> UTCTime
$sel:lastUpdatedTime:DescribeNodeResponse'

-> DescribeNodeResponse

Create a value of DescribeNodeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:assetName:DescribeNodeResponse', describeNodeResponse_assetName - The node's asset name.

$sel:packageArn:DescribeNodeResponse', describeNodeResponse_packageArn - The node's ARN.

$sel:httpStatus:DescribeNodeResponse', describeNodeResponse_httpStatus - The response's http status code.

$sel:nodeId:DescribeNode', describeNodeResponse_nodeId - The node's ID.

$sel:name:DescribeNodeResponse', describeNodeResponse_name - The node's name.

$sel:category:DescribeNodeResponse', describeNodeResponse_category - The node's category.

$sel:ownerAccount:DescribeNode', describeNodeResponse_ownerAccount - The account ID of the node's owner.

$sel:packageName:DescribeNodeResponse', describeNodeResponse_packageName - The node's package name.

$sel:packageId:DescribeNodeResponse', describeNodeResponse_packageId - The node's package ID.

$sel:packageVersion:DescribeNodeResponse', describeNodeResponse_packageVersion - The node's package version.

$sel:patchVersion:DescribeNodeResponse', describeNodeResponse_patchVersion - The node's patch version.

$sel:nodeInterface:DescribeNodeResponse', describeNodeResponse_nodeInterface - The node's interface.

$sel:description:DescribeNodeResponse', describeNodeResponse_description - The node's description.

$sel:createdTime:DescribeNodeResponse', describeNodeResponse_createdTime - When the node was created.

$sel:lastUpdatedTime:DescribeNodeResponse', describeNodeResponse_lastUpdatedTime - When the node was updated.

Response Lenses

describeNodeResponse_assetName :: Lens' DescribeNodeResponse (Maybe Text)

The node's asset name.

describeNodeResponse_packageArn :: Lens' DescribeNodeResponse (Maybe Text)

The node's ARN.

describeNodeResponse_httpStatus :: Lens' DescribeNodeResponse Int

The response's http status code.

describeNodeResponse_nodeId :: Lens' DescribeNodeResponse Text

The node's ID.

describeNodeResponse_name :: Lens' DescribeNodeResponse Text

The node's name.

describeNodeResponse_category :: Lens' DescribeNodeResponse NodeCategory

The node's category.

describeNodeResponse_ownerAccount :: Lens' DescribeNodeResponse Text

The account ID of the node's owner.

describeNodeResponse_packageName :: Lens' DescribeNodeResponse Text

The node's package name.

describeNodeResponse_packageId :: Lens' DescribeNodeResponse Text

The node's package ID.

describeNodeResponse_packageVersion :: Lens' DescribeNodeResponse Text

The node's package version.

describeNodeResponse_patchVersion :: Lens' DescribeNodeResponse Text

The node's patch version.

describeNodeResponse_nodeInterface :: Lens' DescribeNodeResponse NodeInterface

The node's interface.

describeNodeResponse_description :: Lens' DescribeNodeResponse Text

The node's description.

describeNodeResponse_createdTime :: Lens' DescribeNodeResponse UTCTime

When the node was created.

describeNodeResponse_lastUpdatedTime :: Lens' DescribeNodeResponse UTCTime

When the node was updated.