Network.AWS.Panorama.DescribePackageImportJob

Description

Returns information about a package import job.

Creating a Request

data DescribePackageImportJob

See: newDescribePackageImportJob smart constructor.

Constructors

DescribePackageImportJob'

Fields

jobId :: Text

The job's ID.

newDescribePackageImportJob

Arguments

:: Text
$sel:jobId:DescribePackageImportJob'

-> DescribePackageImportJob

Create a value of DescribePackageImportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:jobId:DescribePackageImportJob', describePackageImportJob_jobId - The job's ID.

Request Lenses

describePackageImportJob_jobId :: Lens' DescribePackageImportJob Text

The job's ID.

Destructuring the Response

data DescribePackageImportJobResponse

See: newDescribePackageImportJobResponse smart constructor.

Constructors

DescribePackageImportJobResponse'

Fields

clientToken :: Maybe Text

The job's client token.

jobTags :: Maybe [JobResourceTags]

The job's tags.

httpStatus :: Int

The response's http status code.

jobId :: Text

The job's ID.

jobType :: PackageImportJobType

The job's type.

inputConfig :: PackageImportJobInputConfig

The job's input config.

outputConfig :: PackageImportJobOutputConfig

The job's output config.

output :: PackageImportJobOutput

The job's output.

createdTime :: POSIX

When the job was created.

lastUpdatedTime :: POSIX

When the job was updated.

status :: PackageImportJobStatus

The job's status.

statusMessage :: Text

The job's status message.

newDescribePackageImportJobResponse

Arguments

:: Int
$sel:httpStatus:DescribePackageImportJobResponse'

-> Text
$sel:jobId:DescribePackageImportJob'

-> PackageImportJobType
$sel:jobType:DescribePackageImportJobResponse'

-> PackageImportJobInputConfig
$sel:inputConfig:DescribePackageImportJobResponse'

-> PackageImportJobOutputConfig
$sel:outputConfig:DescribePackageImportJobResponse'

-> PackageImportJobOutput
$sel:output:DescribePackageImportJobResponse'

-> UTCTime
$sel:createdTime:DescribePackageImportJobResponse'

-> UTCTime
$sel:lastUpdatedTime:DescribePackageImportJobResponse'

-> PackageImportJobStatus
$sel:status:DescribePackageImportJobResponse'

-> Text
$sel:statusMessage:DescribePackageImportJobResponse'

-> DescribePackageImportJobResponse

Create a value of DescribePackageImportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:DescribePackageImportJobResponse', describePackageImportJobResponse_clientToken - The job's client token.

$sel:jobTags:DescribePackageImportJobResponse', describePackageImportJobResponse_jobTags - The job's tags.

$sel:httpStatus:DescribePackageImportJobResponse', describePackageImportJobResponse_httpStatus - The response's http status code.

$sel:jobId:DescribePackageImportJob', describePackageImportJobResponse_jobId - The job's ID.

$sel:jobType:DescribePackageImportJobResponse', describePackageImportJobResponse_jobType - The job's type.

$sel:inputConfig:DescribePackageImportJobResponse', describePackageImportJobResponse_inputConfig - The job's input config.

$sel:outputConfig:DescribePackageImportJobResponse', describePackageImportJobResponse_outputConfig - The job's output config.

$sel:output:DescribePackageImportJobResponse', describePackageImportJobResponse_output - The job's output.

$sel:createdTime:DescribePackageImportJobResponse', describePackageImportJobResponse_createdTime - When the job was created.

$sel:lastUpdatedTime:DescribePackageImportJobResponse', describePackageImportJobResponse_lastUpdatedTime - When the job was updated.

$sel:status:DescribePackageImportJobResponse', describePackageImportJobResponse_status - The job's status.

$sel:statusMessage:DescribePackageImportJobResponse', describePackageImportJobResponse_statusMessage - The job's status message.

Response Lenses

describePackageImportJobResponse_clientToken :: Lens' DescribePackageImportJobResponse (Maybe Text)

The job's client token.

describePackageImportJobResponse_jobTags :: Lens' DescribePackageImportJobResponse (Maybe [JobResourceTags])

The job's tags.

describePackageImportJobResponse_httpStatus :: Lens' DescribePackageImportJobResponse Int

The response's http status code.

describePackageImportJobResponse_jobId :: Lens' DescribePackageImportJobResponse Text

The job's ID.

describePackageImportJobResponse_jobType :: Lens' DescribePackageImportJobResponse PackageImportJobType

The job's type.

describePackageImportJobResponse_inputConfig :: Lens' DescribePackageImportJobResponse PackageImportJobInputConfig

The job's input config.

describePackageImportJobResponse_outputConfig :: Lens' DescribePackageImportJobResponse PackageImportJobOutputConfig

The job's output config.

describePackageImportJobResponse_output :: Lens' DescribePackageImportJobResponse PackageImportJobOutput

The job's output.

describePackageImportJobResponse_createdTime :: Lens' DescribePackageImportJobResponse UTCTime

When the job was created.

describePackageImportJobResponse_lastUpdatedTime :: Lens' DescribePackageImportJobResponse UTCTime

When the job was updated.

describePackageImportJobResponse_status :: Lens' DescribePackageImportJobResponse PackageImportJobStatus

The job's status.

describePackageImportJobResponse_statusMessage :: Lens' DescribePackageImportJobResponse Text

The job's status message.