Network.AWS.Panorama.ListApplicationInstances

Description

Returns a list of application instances.

Creating a Request

data ListApplicationInstances

See: newListApplicationInstances smart constructor.

Constructors

ListApplicationInstances'

Fields

nextToken :: Maybe Text

Specify the pagination token from a previous request to retrieve the next page of results.

statusFilter :: Maybe StatusFilter

Only include instances with a specific status.

deviceId :: Maybe Text

The application instances' device ID.

maxResults :: Maybe Natural

The maximum number of application instances to return in one page of results.

newListApplicationInstances :: ListApplicationInstances

Create a value of ListApplicationInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListApplicationInstances', listApplicationInstances_nextToken - Specify the pagination token from a previous request to retrieve the next page of results.

$sel:statusFilter:ListApplicationInstances', listApplicationInstances_statusFilter - Only include instances with a specific status.

$sel:deviceId:ListApplicationInstances', listApplicationInstances_deviceId - The application instances' device ID.

$sel:maxResults:ListApplicationInstances', listApplicationInstances_maxResults - The maximum number of application instances to return in one page of results.

Request Lenses

listApplicationInstances_nextToken :: Lens' ListApplicationInstances (Maybe Text)

Specify the pagination token from a previous request to retrieve the next page of results.

listApplicationInstances_statusFilter :: Lens' ListApplicationInstances (Maybe StatusFilter)

Only include instances with a specific status.

listApplicationInstances_deviceId :: Lens' ListApplicationInstances (Maybe Text)

The application instances' device ID.

listApplicationInstances_maxResults :: Lens' ListApplicationInstances (Maybe Natural)

The maximum number of application instances to return in one page of results.

Destructuring the Response

data ListApplicationInstancesResponse

See: newListApplicationInstancesResponse smart constructor.

Constructors

ListApplicationInstancesResponse'

Fields

nextToken :: Maybe Text

A pagination token that's included if more results are available.

applicationInstances :: Maybe [ApplicationInstance]

A list of application instances.

httpStatus :: Int

The response's http status code.

newListApplicationInstancesResponse

Arguments

:: Int
$sel:httpStatus:ListApplicationInstancesResponse'

-> ListApplicationInstancesResponse

Create a value of ListApplicationInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListApplicationInstances', listApplicationInstancesResponse_nextToken - A pagination token that's included if more results are available.

$sel:applicationInstances:ListApplicationInstancesResponse', listApplicationInstancesResponse_applicationInstances - A list of application instances.

$sel:httpStatus:ListApplicationInstancesResponse', listApplicationInstancesResponse_httpStatus - The response's http status code.

Response Lenses

listApplicationInstancesResponse_nextToken :: Lens' ListApplicationInstancesResponse (Maybe Text)

A pagination token that's included if more results are available.

listApplicationInstancesResponse_applicationInstances :: Lens' ListApplicationInstancesResponse (Maybe [ApplicationInstance])

A list of application instances.

listApplicationInstancesResponse_httpStatus :: Lens' ListApplicationInstancesResponse Int

The response's http status code.