Network.AWS.Panorama.ListNodes

Description

Returns a list of nodes.

Creating a Request

data ListNodes

See: newListNodes smart constructor.

Constructors

ListNodes'

Fields

patchVersion :: Maybe Text

Search for nodes by patch version.

packageName :: Maybe Text

Search for nodes by name.

packageVersion :: Maybe Text

Search for nodes by version.

category :: Maybe NodeCategory

Search for nodes by category.

nextToken :: Maybe Text

Specify the pagination token from a previous request to retrieve the next page of results.

ownerAccount :: Maybe Text

Search for nodes by the account ID of the nodes' owner.

maxResults :: Maybe Natural

The maximum number of nodes to return in one page of results.

newListNodes :: ListNodes

Create a value of ListNodes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:patchVersion:ListNodes', listNodes_patchVersion - Search for nodes by patch version.

$sel:packageName:ListNodes', listNodes_packageName - Search for nodes by name.

$sel:packageVersion:ListNodes', listNodes_packageVersion - Search for nodes by version.

$sel:category:ListNodes', listNodes_category - Search for nodes by category.

$sel:nextToken:ListNodes', listNodes_nextToken - Specify the pagination token from a previous request to retrieve the next page of results.

$sel:ownerAccount:ListNodes', listNodes_ownerAccount - Search for nodes by the account ID of the nodes' owner.

$sel:maxResults:ListNodes', listNodes_maxResults - The maximum number of nodes to return in one page of results.

Request Lenses

listNodes_patchVersion :: Lens' ListNodes (Maybe Text)

Search for nodes by patch version.

listNodes_packageName :: Lens' ListNodes (Maybe Text)

Search for nodes by name.

listNodes_packageVersion :: Lens' ListNodes (Maybe Text)

Search for nodes by version.

listNodes_category :: Lens' ListNodes (Maybe NodeCategory)

Search for nodes by category.

listNodes_nextToken :: Lens' ListNodes (Maybe Text)

Specify the pagination token from a previous request to retrieve the next page of results.

listNodes_ownerAccount :: Lens' ListNodes (Maybe Text)

Search for nodes by the account ID of the nodes' owner.

listNodes_maxResults :: Lens' ListNodes (Maybe Natural)

The maximum number of nodes to return in one page of results.

Destructuring the Response

data ListNodesResponse

See: newListNodesResponse smart constructor.

Constructors

ListNodesResponse'

Fields

nextToken :: Maybe Text

A pagination token that's included if more results are available.

nodes :: Maybe [Node]

A list of nodes.

httpStatus :: Int

The response's http status code.

newListNodesResponse

Arguments

:: Int
$sel:httpStatus:ListNodesResponse'

-> ListNodesResponse

Create a value of ListNodesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListNodes', listNodesResponse_nextToken - A pagination token that's included if more results are available.

$sel:nodes:ListNodesResponse', listNodesResponse_nodes - A list of nodes.

$sel:httpStatus:ListNodesResponse', listNodesResponse_httpStatus - The response's http status code.

Response Lenses

listNodesResponse_nextToken :: Lens' ListNodesResponse (Maybe Text)

A pagination token that's included if more results are available.

listNodesResponse_nodes :: Lens' ListNodesResponse (Maybe [Node])

A list of nodes.

listNodesResponse_httpStatus :: Lens' ListNodesResponse Int

The response's http status code.