Network.AWS.Panorama.ProvisionDevice

Description

Creates a device and returns a configuration archive. The configuration archive is a ZIP file that contains a provisioning certificate that is valid for 5 minutes. Transfer the configuration archive to the device with the included USB storage device within 5 minutes.

Creating a Request

data ProvisionDevice

See: newProvisionDevice smart constructor.

Constructors

ProvisionDevice'

Fields

description :: Maybe Text

A description for the device.

networkingConfiguration :: Maybe NetworkPayload

A networking configuration for the device.

tags :: Maybe (HashMap Text Text)

Tags for the device.

name :: Text

A name for the device.

newProvisionDevice

Arguments

:: Text
$sel:name:ProvisionDevice'

-> ProvisionDevice

Create a value of ProvisionDevice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:ProvisionDevice', provisionDevice_description - A description for the device.

$sel:networkingConfiguration:ProvisionDevice', provisionDevice_networkingConfiguration - A networking configuration for the device.

$sel:tags:ProvisionDevice', provisionDevice_tags - Tags for the device.

$sel:name:ProvisionDevice', provisionDevice_name - A name for the device.

Request Lenses

provisionDevice_description :: Lens' ProvisionDevice (Maybe Text)

A description for the device.

provisionDevice_networkingConfiguration :: Lens' ProvisionDevice (Maybe NetworkPayload)

A networking configuration for the device.

provisionDevice_tags :: Lens' ProvisionDevice (Maybe (HashMap Text Text))

Tags for the device.

provisionDevice_name :: Lens' ProvisionDevice Text

A name for the device.

Destructuring the Response

data ProvisionDeviceResponse

See: newProvisionDeviceResponse smart constructor.

Constructors

ProvisionDeviceResponse'

Fields

certificates :: Maybe Base64

The device's configuration bundle.

deviceId :: Maybe Text

The device's ID.

iotThingName :: Maybe Text

The device's IoT thing name.

httpStatus :: Int

The response's http status code.

arn :: Text

The device's ARN.

status :: DeviceStatus

The device's status.

newProvisionDeviceResponse

Arguments

:: Int
$sel:httpStatus:ProvisionDeviceResponse'

-> Text
$sel:arn:ProvisionDeviceResponse'

-> DeviceStatus
$sel:status:ProvisionDeviceResponse'

-> ProvisionDeviceResponse

Create a value of ProvisionDeviceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificates:ProvisionDeviceResponse', provisionDeviceResponse_certificates - The device's configuration bundle.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:deviceId:ProvisionDeviceResponse', provisionDeviceResponse_deviceId - The device's ID.

$sel:iotThingName:ProvisionDeviceResponse', provisionDeviceResponse_iotThingName - The device's IoT thing name.

$sel:httpStatus:ProvisionDeviceResponse', provisionDeviceResponse_httpStatus - The response's http status code.

$sel:arn:ProvisionDeviceResponse', provisionDeviceResponse_arn - The device's ARN.

$sel:status:ProvisionDeviceResponse', provisionDeviceResponse_status - The device's status.

Response Lenses

provisionDeviceResponse_certificates :: Lens' ProvisionDeviceResponse (Maybe ByteString)

The device's configuration bundle.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

provisionDeviceResponse_deviceId :: Lens' ProvisionDeviceResponse (Maybe Text)

The device's ID.

provisionDeviceResponse_iotThingName :: Lens' ProvisionDeviceResponse (Maybe Text)

The device's IoT thing name.

provisionDeviceResponse_httpStatus :: Lens' ProvisionDeviceResponse Int

The response's http status code.

provisionDeviceResponse_arn :: Lens' ProvisionDeviceResponse Text

The device's ARN.

provisionDeviceResponse_status :: Lens' ProvisionDeviceResponse DeviceStatus

The device's status.