Network.AWS.Panorama.Types.Node

Description

Documentation

data Node

An application node that represents a camera stream, a model, code, or output.

See: newNode smart constructor.

Constructors

Node'

Fields

packageArn :: Maybe Text

The node's ARN.

ownerAccount :: Maybe Text

The account ID of the node's owner.

description :: Maybe Text

The node's description.

nodeId :: Text

The node's ID.

name :: Text

The node's name.

category :: NodeCategory

The node's category.

packageName :: Text

The node's package name.

packageId :: Text

The node's package ID.

packageVersion :: Text

The node's package version.

patchVersion :: Text

The node's patch version.

createdTime :: POSIX

When the node was created.

newNode

Arguments

:: Text
$sel:nodeId:Node'

-> Text
$sel:name:Node'

-> NodeCategory
$sel:category:Node'

-> Text
$sel:packageName:Node'

-> Text
$sel:packageId:Node'

-> Text
$sel:packageVersion:Node'

-> Text
$sel:patchVersion:Node'

-> UTCTime
$sel:createdTime:Node'

-> Node

Create a value of Node with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:packageArn:Node', node_packageArn - The node's ARN.

$sel:ownerAccount:Node', node_ownerAccount - The account ID of the node's owner.

$sel:description:Node', node_description - The node's description.

$sel:nodeId:Node', node_nodeId - The node's ID.

$sel:name:Node', node_name - The node's name.

$sel:category:Node', node_category - The node's category.

$sel:packageName:Node', node_packageName - The node's package name.

$sel:packageId:Node', node_packageId - The node's package ID.

$sel:packageVersion:Node', node_packageVersion - The node's package version.

$sel:patchVersion:Node', node_patchVersion - The node's patch version.

$sel:createdTime:Node', node_createdTime - When the node was created.

node_packageArn :: Lens' Node (Maybe Text)

The node's ARN.

node_ownerAccount :: Lens' Node (Maybe Text)

The account ID of the node's owner.

node_description :: Lens' Node (Maybe Text)

The node's description.

node_nodeId :: Lens' Node Text

The node's ID.

node_name :: Lens' Node Text

The node's name.

node_category :: Lens' Node NodeCategory

The node's category.

node_packageName :: Lens' Node Text

The node's package name.

node_packageId :: Lens' Node Text

The node's package ID.

node_packageVersion :: Lens' Node Text

The node's package version.

node_patchVersion :: Lens' Node Text

The node's patch version.

node_createdTime :: Lens' Node UTCTime

When the node was created.