Network.AWS.Panorama.Types.PackageVersionOutputConfig

Description

Documentation

data PackageVersionOutputConfig

A package version output configuration.

See: newPackageVersionOutputConfig smart constructor.

Constructors

PackageVersionOutputConfig'

Fields

markLatest :: Maybe Bool

Indicates that the version is recommended for all users.

packageName :: Text

The output's package name.

packageVersion :: Text

The output's package version.

newPackageVersionOutputConfig

Arguments

:: Text
$sel:packageName:PackageVersionOutputConfig'

-> Text
$sel:packageVersion:PackageVersionOutputConfig'

-> PackageVersionOutputConfig

Create a value of PackageVersionOutputConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:markLatest:PackageVersionOutputConfig', packageVersionOutputConfig_markLatest - Indicates that the version is recommended for all users.

$sel:packageName:PackageVersionOutputConfig', packageVersionOutputConfig_packageName - The output's package name.

$sel:packageVersion:PackageVersionOutputConfig', packageVersionOutputConfig_packageVersion - The output's package version.

packageVersionOutputConfig_markLatest :: Lens' PackageVersionOutputConfig (Maybe Bool)

Indicates that the version is recommended for all users.

packageVersionOutputConfig_packageName :: Lens' PackageVersionOutputConfig Text

The output's package name.

packageVersionOutputConfig_packageVersion :: Lens' PackageVersionOutputConfig Text

The output's package version.