Network.AWS.Personalize.CreateBatchInferenceJob

Description

Creates a batch inference job. The operation can handle up to 50 million records and the input file must be in JSON format. For more information, see recommendations-batch.

Creating a Request

data CreateBatchInferenceJob

See: newCreateBatchInferenceJob smart constructor.

Constructors

CreateBatchInferenceJob'

Fields

numResults :: Maybe Int

The number of recommendations to retreive.

batchInferenceJobConfig :: Maybe BatchInferenceJobConfig

The configuration details of a batch inference job.

filterArn :: Maybe Text

The ARN of the filter to apply to the batch inference job. For more information on using filters, see Filtering Batch Recommendations..

jobName :: Text

The name of the batch inference job to create.

solutionVersionArn :: Text

The Amazon Resource Name (ARN) of the solution version that will be used to generate the batch inference recommendations.

jobInput :: BatchInferenceJobInput

The Amazon S3 path that leads to the input file to base your recommendations on. The input material must be in JSON format.

jobOutput :: BatchInferenceJobOutput

The path to the Amazon S3 bucket where the job's output will be stored.

roleArn :: Text

The ARN of the Amazon Identity and Access Management role that has permissions to read and write to your input and output Amazon S3 buckets respectively.

newCreateBatchInferenceJob

Arguments

:: Text
$sel:jobName:CreateBatchInferenceJob'

-> Text
$sel:solutionVersionArn:CreateBatchInferenceJob'

-> BatchInferenceJobInput
$sel:jobInput:CreateBatchInferenceJob'

-> BatchInferenceJobOutput
$sel:jobOutput:CreateBatchInferenceJob'

-> Text
$sel:roleArn:CreateBatchInferenceJob'

-> CreateBatchInferenceJob

Create a value of CreateBatchInferenceJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:numResults:CreateBatchInferenceJob', createBatchInferenceJob_numResults - The number of recommendations to retreive.

$sel:batchInferenceJobConfig:CreateBatchInferenceJob', createBatchInferenceJob_batchInferenceJobConfig - The configuration details of a batch inference job.

$sel:filterArn:CreateBatchInferenceJob', createBatchInferenceJob_filterArn - The ARN of the filter to apply to the batch inference job. For more information on using filters, see Filtering Batch Recommendations..

$sel:jobName:CreateBatchInferenceJob', createBatchInferenceJob_jobName - The name of the batch inference job to create.

$sel:solutionVersionArn:CreateBatchInferenceJob', createBatchInferenceJob_solutionVersionArn - The Amazon Resource Name (ARN) of the solution version that will be used to generate the batch inference recommendations.

$sel:jobInput:CreateBatchInferenceJob', createBatchInferenceJob_jobInput - The Amazon S3 path that leads to the input file to base your recommendations on. The input material must be in JSON format.

$sel:jobOutput:CreateBatchInferenceJob', createBatchInferenceJob_jobOutput - The path to the Amazon S3 bucket where the job's output will be stored.

$sel:roleArn:CreateBatchInferenceJob', createBatchInferenceJob_roleArn - The ARN of the Amazon Identity and Access Management role that has permissions to read and write to your input and output Amazon S3 buckets respectively.

Request Lenses

createBatchInferenceJob_numResults :: Lens' CreateBatchInferenceJob (Maybe Int)

The number of recommendations to retreive.

createBatchInferenceJob_batchInferenceJobConfig :: Lens' CreateBatchInferenceJob (Maybe BatchInferenceJobConfig)

The configuration details of a batch inference job.

createBatchInferenceJob_filterArn :: Lens' CreateBatchInferenceJob (Maybe Text)

The ARN of the filter to apply to the batch inference job. For more information on using filters, see Filtering Batch Recommendations..

createBatchInferenceJob_jobName :: Lens' CreateBatchInferenceJob Text

The name of the batch inference job to create.

createBatchInferenceJob_solutionVersionArn :: Lens' CreateBatchInferenceJob Text

The Amazon Resource Name (ARN) of the solution version that will be used to generate the batch inference recommendations.

createBatchInferenceJob_jobInput :: Lens' CreateBatchInferenceJob BatchInferenceJobInput

The Amazon S3 path that leads to the input file to base your recommendations on. The input material must be in JSON format.

createBatchInferenceJob_jobOutput :: Lens' CreateBatchInferenceJob BatchInferenceJobOutput

The path to the Amazon S3 bucket where the job's output will be stored.

createBatchInferenceJob_roleArn :: Lens' CreateBatchInferenceJob Text

The ARN of the Amazon Identity and Access Management role that has permissions to read and write to your input and output Amazon S3 buckets respectively.

Destructuring the Response

data CreateBatchInferenceJobResponse

See: newCreateBatchInferenceJobResponse smart constructor.

Constructors

CreateBatchInferenceJobResponse'

Fields

batchInferenceJobArn :: Maybe Text

The ARN of the batch inference job.

httpStatus :: Int

The response's http status code.

newCreateBatchInferenceJobResponse

Arguments

:: Int
$sel:httpStatus:CreateBatchInferenceJobResponse'

-> CreateBatchInferenceJobResponse

Create a value of CreateBatchInferenceJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:batchInferenceJobArn:CreateBatchInferenceJobResponse', createBatchInferenceJobResponse_batchInferenceJobArn - The ARN of the batch inference job.

$sel:httpStatus:CreateBatchInferenceJobResponse', createBatchInferenceJobResponse_httpStatus - The response's http status code.

Response Lenses

createBatchInferenceJobResponse_batchInferenceJobArn :: Lens' CreateBatchInferenceJobResponse (Maybe Text)

The ARN of the batch inference job.

createBatchInferenceJobResponse_httpStatus :: Lens' CreateBatchInferenceJobResponse Int

The response's http status code.