Network.AWS.Personalize.CreateDatasetExportJob

Description

Creates a job that exports data from your dataset to an Amazon S3 bucket. To allow Amazon Personalize to export the training data, you must specify an service-linked IAM role that gives Amazon Personalize PutObject permissions for your Amazon S3 bucket. For information, see Exporting a dataset in the Amazon Personalize developer guide.

Status

A dataset export job can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

To get the status of the export job, call DescribeDatasetExportJob, and specify the Amazon Resource Name (ARN) of the dataset export job. The dataset export is complete when the status shows as ACTIVE. If the status shows as CREATE FAILED, the response includes a failureReason key, which describes why the job failed.

Creating a Request

data CreateDatasetExportJob

See: newCreateDatasetExportJob smart constructor.

Constructors

CreateDatasetExportJob'

Fields

ingestionMode :: Maybe IngestionMode

The data to export, based on how you imported the data. You can choose to export only BULK data that you imported using a dataset import job, only PUT data that you imported incrementally (using the console, PutEvents, PutUsers and PutItems operations), or ALL for both types. The default value is PUT.

jobName :: Text

The name for the dataset export job.

datasetArn :: Text

The Amazon Resource Name (ARN) of the dataset that contains the data to export.

roleArn :: Text

The Amazon Resource Name (ARN) of the IAM service role that has permissions to add data to your output Amazon S3 bucket.

jobOutput :: DatasetExportJobOutput

The path to the Amazon S3 bucket where the job's output is stored.

newCreateDatasetExportJob

Arguments

:: Text
$sel:jobName:CreateDatasetExportJob'

-> Text
$sel:datasetArn:CreateDatasetExportJob'

-> Text
$sel:roleArn:CreateDatasetExportJob'

-> DatasetExportJobOutput
$sel:jobOutput:CreateDatasetExportJob'

-> CreateDatasetExportJob

Create a value of CreateDatasetExportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ingestionMode:CreateDatasetExportJob', createDatasetExportJob_ingestionMode - The data to export, based on how you imported the data. You can choose to export only BULK data that you imported using a dataset import job, only PUT data that you imported incrementally (using the console, PutEvents, PutUsers and PutItems operations), or ALL for both types. The default value is PUT.

$sel:jobName:CreateDatasetExportJob', createDatasetExportJob_jobName - The name for the dataset export job.

$sel:datasetArn:CreateDatasetExportJob', createDatasetExportJob_datasetArn - The Amazon Resource Name (ARN) of the dataset that contains the data to export.

$sel:roleArn:CreateDatasetExportJob', createDatasetExportJob_roleArn - The Amazon Resource Name (ARN) of the IAM service role that has permissions to add data to your output Amazon S3 bucket.

$sel:jobOutput:CreateDatasetExportJob', createDatasetExportJob_jobOutput - The path to the Amazon S3 bucket where the job's output is stored.

Request Lenses

createDatasetExportJob_ingestionMode :: Lens' CreateDatasetExportJob (Maybe IngestionMode)

The data to export, based on how you imported the data. You can choose to export only BULK data that you imported using a dataset import job, only PUT data that you imported incrementally (using the console, PutEvents, PutUsers and PutItems operations), or ALL for both types. The default value is PUT.

createDatasetExportJob_jobName :: Lens' CreateDatasetExportJob Text

The name for the dataset export job.

createDatasetExportJob_datasetArn :: Lens' CreateDatasetExportJob Text

The Amazon Resource Name (ARN) of the dataset that contains the data to export.

createDatasetExportJob_roleArn :: Lens' CreateDatasetExportJob Text

The Amazon Resource Name (ARN) of the IAM service role that has permissions to add data to your output Amazon S3 bucket.

createDatasetExportJob_jobOutput :: Lens' CreateDatasetExportJob DatasetExportJobOutput

The path to the Amazon S3 bucket where the job's output is stored.

Destructuring the Response

data CreateDatasetExportJobResponse

See: newCreateDatasetExportJobResponse smart constructor.

Constructors

CreateDatasetExportJobResponse'

Fields

datasetExportJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset export job.

httpStatus :: Int

The response's http status code.

newCreateDatasetExportJobResponse

Arguments

:: Int
$sel:httpStatus:CreateDatasetExportJobResponse'

-> CreateDatasetExportJobResponse

Create a value of CreateDatasetExportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetExportJobArn:CreateDatasetExportJobResponse', createDatasetExportJobResponse_datasetExportJobArn - The Amazon Resource Name (ARN) of the dataset export job.

$sel:httpStatus:CreateDatasetExportJobResponse', createDatasetExportJobResponse_httpStatus - The response's http status code.

Response Lenses

createDatasetExportJobResponse_datasetExportJobArn :: Lens' CreateDatasetExportJobResponse (Maybe Text)

The Amazon Resource Name (ARN) of the dataset export job.

createDatasetExportJobResponse_httpStatus :: Lens' CreateDatasetExportJobResponse Int

The response's http status code.