Network.AWS.Personalize.CreateEventTracker

Description

Creates an event tracker that you use when adding event data to a specified dataset group using the PutEvents API.

Only one event tracker can be associated with a dataset group. You will get an error if you call CreateEventTracker using the same dataset group as an existing event tracker.

When you create an event tracker, the response includes a tracking ID, which you pass as a parameter when you use the PutEvents operation. Amazon Personalize then appends the event data to the Interactions dataset of the dataset group you specify in your event tracker.

The event tracker can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

DELETE PENDING > DELETE IN_PROGRESS

To get the status of the event tracker, call DescribeEventTracker.

The event tracker must be in the ACTIVE state before using the tracking ID.

Related APIs

ListEventTrackers

DescribeEventTracker

DeleteEventTracker

Creating a Request

data CreateEventTracker

See: newCreateEventTracker smart constructor.

Constructors

CreateEventTracker'

Fields

name :: Text

The name for the event tracker.

datasetGroupArn :: Text

The Amazon Resource Name (ARN) of the dataset group that receives the event data.

newCreateEventTracker

Arguments

:: Text
$sel:name:CreateEventTracker'

-> Text
$sel:datasetGroupArn:CreateEventTracker'

-> CreateEventTracker

Create a value of CreateEventTracker with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:CreateEventTracker', createEventTracker_name - The name for the event tracker.

$sel:datasetGroupArn:CreateEventTracker', createEventTracker_datasetGroupArn - The Amazon Resource Name (ARN) of the dataset group that receives the event data.

Request Lenses

createEventTracker_name :: Lens' CreateEventTracker Text

The name for the event tracker.

createEventTracker_datasetGroupArn :: Lens' CreateEventTracker Text

The Amazon Resource Name (ARN) of the dataset group that receives the event data.

Destructuring the Response

data CreateEventTrackerResponse

See: newCreateEventTrackerResponse smart constructor.

Constructors

CreateEventTrackerResponse'

Fields

trackingId :: Maybe Text

The ID of the event tracker. Include this ID in requests to the PutEvents API.

eventTrackerArn :: Maybe Text

The ARN of the event tracker.

httpStatus :: Int

The response's http status code.

newCreateEventTrackerResponse

Arguments

:: Int
$sel:httpStatus:CreateEventTrackerResponse'

-> CreateEventTrackerResponse

Create a value of CreateEventTrackerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trackingId:CreateEventTrackerResponse', createEventTrackerResponse_trackingId - The ID of the event tracker. Include this ID in requests to the PutEvents API.

$sel:eventTrackerArn:CreateEventTrackerResponse', createEventTrackerResponse_eventTrackerArn - The ARN of the event tracker.

$sel:httpStatus:CreateEventTrackerResponse', createEventTrackerResponse_httpStatus - The response's http status code.

Response Lenses

createEventTrackerResponse_trackingId :: Lens' CreateEventTrackerResponse (Maybe Text)

The ID of the event tracker. Include this ID in requests to the PutEvents API.

createEventTrackerResponse_eventTrackerArn :: Lens' CreateEventTrackerResponse (Maybe Text)

The ARN of the event tracker.

createEventTrackerResponse_httpStatus :: Lens' CreateEventTrackerResponse Int

The response's http status code.