Network.AWS.Personalize.CreateSolution

Description

Creates the configuration for training a model. A trained model is known as a solution. After the configuration is created, you train the model (create a solution) by calling the CreateSolutionVersion operation. Every time you call CreateSolutionVersion, a new version of the solution is created.

After creating a solution version, you check its accuracy by calling GetSolutionMetrics. When you are satisfied with the version, you deploy it using CreateCampaign. The campaign provides recommendations to a client through the GetRecommendations API.

To train a model, Amazon Personalize requires training data and a recipe. The training data comes from the dataset group that you provide in the request. A recipe specifies the training algorithm and a feature transformation. You can specify one of the predefined recipes provided by Amazon Personalize. Alternatively, you can specify performAutoML and Amazon Personalize will analyze your data and select the optimum USER_PERSONALIZATION recipe for you.

Amazon Personalize doesn't support configuring the hpoObjective for solution hyperparameter optimization at this time.

Status

A solution can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

DELETE PENDING > DELETE IN_PROGRESS

To get the status of the solution, call DescribeSolution. Wait until the status shows as ACTIVE before calling CreateSolutionVersion.

Related APIs

ListSolutions

CreateSolutionVersion

DescribeSolution

DeleteSolution

ListSolutionVersions

DescribeSolutionVersion

Creating a Request

data CreateSolution

See: newCreateSolution smart constructor.

Constructors

CreateSolution'

Fields

performAutoML :: Maybe Bool

Whether to perform automated machine learning (AutoML). The default is false. For this case, you must specify recipeArn.

When set to true, Amazon Personalize analyzes your training data and selects the optimal USER_PERSONALIZATION recipe and hyperparameters. In this case, you must omit recipeArn. Amazon Personalize determines the optimal recipe by running tests with different values for the hyperparameters. AutoML lengthens the training process as compared to selecting a specific recipe.

recipeArn :: Maybe Text

The ARN of the recipe to use for model training. Only specified when performAutoML is false.

eventType :: Maybe Text

When your have multiple event types (using an EVENT_TYPE schema field), this parameter specifies which event type (for example, 'click' or 'like') is used for training the model.

If you do not provide an eventType, Amazon Personalize will use all interactions for training with equal weight regardless of type.

solutionConfig :: Maybe SolutionConfig

The configuration to use with the solution. When performAutoML is set to true, Amazon Personalize only evaluates the autoMLConfig section of the solution configuration.

Amazon Personalize doesn't support configuring the hpoObjective at this time.

performHPO :: Maybe Bool

Whether to perform hyperparameter optimization (HPO) on the specified or selected recipe. The default is false.

When performing AutoML, this parameter is always true and you should not set it to false.

name :: Text

The name for the solution.

datasetGroupArn :: Text

The Amazon Resource Name (ARN) of the dataset group that provides the training data.

newCreateSolution

Arguments

:: Text
$sel:name:CreateSolution'

-> Text
$sel:datasetGroupArn:CreateSolution'

-> CreateSolution

Create a value of CreateSolution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:performAutoML:CreateSolution', createSolution_performAutoML - Whether to perform automated machine learning (AutoML). The default is false. For this case, you must specify recipeArn.

When set to true, Amazon Personalize analyzes your training data and selects the optimal USER_PERSONALIZATION recipe and hyperparameters. In this case, you must omit recipeArn. Amazon Personalize determines the optimal recipe by running tests with different values for the hyperparameters. AutoML lengthens the training process as compared to selecting a specific recipe.

$sel:recipeArn:CreateSolution', createSolution_recipeArn - The ARN of the recipe to use for model training. Only specified when performAutoML is false.

$sel:eventType:CreateSolution', createSolution_eventType - When your have multiple event types (using an EVENT_TYPE schema field), this parameter specifies which event type (for example, 'click' or 'like') is used for training the model.

If you do not provide an eventType, Amazon Personalize will use all interactions for training with equal weight regardless of type.

$sel:solutionConfig:CreateSolution', createSolution_solutionConfig - The configuration to use with the solution. When performAutoML is set to true, Amazon Personalize only evaluates the autoMLConfig section of the solution configuration.

Amazon Personalize doesn't support configuring the hpoObjective at this time.

$sel:performHPO:CreateSolution', createSolution_performHPO - Whether to perform hyperparameter optimization (HPO) on the specified or selected recipe. The default is false.

When performing AutoML, this parameter is always true and you should not set it to false.

$sel:name:CreateSolution', createSolution_name - The name for the solution.

$sel:datasetGroupArn:CreateSolution', createSolution_datasetGroupArn - The Amazon Resource Name (ARN) of the dataset group that provides the training data.

Request Lenses

createSolution_performAutoML :: Lens' CreateSolution (Maybe Bool)

Whether to perform automated machine learning (AutoML). The default is false. For this case, you must specify recipeArn.

When set to true, Amazon Personalize analyzes your training data and selects the optimal USER_PERSONALIZATION recipe and hyperparameters. In this case, you must omit recipeArn. Amazon Personalize determines the optimal recipe by running tests with different values for the hyperparameters. AutoML lengthens the training process as compared to selecting a specific recipe.

createSolution_recipeArn :: Lens' CreateSolution (Maybe Text)

The ARN of the recipe to use for model training. Only specified when performAutoML is false.

createSolution_eventType :: Lens' CreateSolution (Maybe Text)

When your have multiple event types (using an EVENT_TYPE schema field), this parameter specifies which event type (for example, 'click' or 'like') is used for training the model.

If you do not provide an eventType, Amazon Personalize will use all interactions for training with equal weight regardless of type.

createSolution_solutionConfig :: Lens' CreateSolution (Maybe SolutionConfig)

The configuration to use with the solution. When performAutoML is set to true, Amazon Personalize only evaluates the autoMLConfig section of the solution configuration.

Amazon Personalize doesn't support configuring the hpoObjective at this time.

createSolution_performHPO :: Lens' CreateSolution (Maybe Bool)

Whether to perform hyperparameter optimization (HPO) on the specified or selected recipe. The default is false.

When performing AutoML, this parameter is always true and you should not set it to false.

createSolution_name :: Lens' CreateSolution Text

The name for the solution.

createSolution_datasetGroupArn :: Lens' CreateSolution Text

The Amazon Resource Name (ARN) of the dataset group that provides the training data.

Destructuring the Response

data CreateSolutionResponse

See: newCreateSolutionResponse smart constructor.

Constructors

CreateSolutionResponse'

Fields

solutionArn :: Maybe Text

The ARN of the solution.

httpStatus :: Int

The response's http status code.

newCreateSolutionResponse

Arguments

:: Int
$sel:httpStatus:CreateSolutionResponse'

-> CreateSolutionResponse

Create a value of CreateSolutionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:solutionArn:CreateSolutionResponse', createSolutionResponse_solutionArn - The ARN of the solution.

$sel:httpStatus:CreateSolutionResponse', createSolutionResponse_httpStatus - The response's http status code.

Response Lenses

createSolutionResponse_solutionArn :: Lens' CreateSolutionResponse (Maybe Text)

The ARN of the solution.

createSolutionResponse_httpStatus :: Lens' CreateSolutionResponse Int

The response's http status code.