Network.AWS.Personalize.CreateSolutionVersion

Description

Trains or retrains an active solution. A solution is created using the CreateSolution operation and must be in the ACTIVE state before calling CreateSolutionVersion. A new version of the solution is created every time you call this operation.

Status

A solution version can be in one of the following states:

CREATE PENDING

CREATE IN_PROGRESS

ACTIVE

CREATE FAILED

CREATE STOPPING

CREATE STOPPED

To get the status of the version, call DescribeSolutionVersion. Wait until the status shows as ACTIVE before calling CreateCampaign.

If the status shows as CREATE FAILED, the response includes a failureReason key, which describes why the job failed.

Related APIs

ListSolutionVersions

DescribeSolutionVersion

ListSolutions

CreateSolution

DescribeSolution

DeleteSolution

Creating a Request

data CreateSolutionVersion

See: newCreateSolutionVersion smart constructor.

Constructors

CreateSolutionVersion'

Fields

trainingMode :: Maybe TrainingMode

The scope of training to be performed when creating the solution version. The FULL option trains the solution version based on the entirety of the input solution's training data, while the UPDATE option processes only the data that has changed in comparison to the input solution. Choose UPDATE when you want to incrementally update your solution version instead of creating an entirely new one.

The UPDATE option can only be used when you already have an active solution version created from the input solution using the FULL option and the input solution was trained with the User-Personalization recipe or the HRNN-Coldstart recipe.

solutionArn :: Text

The Amazon Resource Name (ARN) of the solution containing the training configuration information.

newCreateSolutionVersion

Arguments

:: Text
$sel:solutionArn:CreateSolutionVersion'

-> CreateSolutionVersion

Create a value of CreateSolutionVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trainingMode:CreateSolutionVersion', createSolutionVersion_trainingMode - The scope of training to be performed when creating the solution version. The FULL option trains the solution version based on the entirety of the input solution's training data, while the UPDATE option processes only the data that has changed in comparison to the input solution. Choose UPDATE when you want to incrementally update your solution version instead of creating an entirely new one.

The UPDATE option can only be used when you already have an active solution version created from the input solution using the FULL option and the input solution was trained with the User-Personalization recipe or the HRNN-Coldstart recipe.

$sel:solutionArn:CreateSolutionVersion', createSolutionVersion_solutionArn - The Amazon Resource Name (ARN) of the solution containing the training configuration information.

Request Lenses

createSolutionVersion_trainingMode :: Lens' CreateSolutionVersion (Maybe TrainingMode)

The scope of training to be performed when creating the solution version. The FULL option trains the solution version based on the entirety of the input solution's training data, while the UPDATE option processes only the data that has changed in comparison to the input solution. Choose UPDATE when you want to incrementally update your solution version instead of creating an entirely new one.

The UPDATE option can only be used when you already have an active solution version created from the input solution using the FULL option and the input solution was trained with the User-Personalization recipe or the HRNN-Coldstart recipe.

createSolutionVersion_solutionArn :: Lens' CreateSolutionVersion Text

The Amazon Resource Name (ARN) of the solution containing the training configuration information.

Destructuring the Response

data CreateSolutionVersionResponse

See: newCreateSolutionVersionResponse smart constructor.

Constructors

CreateSolutionVersionResponse'

Fields

solutionVersionArn :: Maybe Text

The ARN of the new solution version.

httpStatus :: Int

The response's http status code.

newCreateSolutionVersionResponse

Arguments

:: Int
$sel:httpStatus:CreateSolutionVersionResponse'

-> CreateSolutionVersionResponse

Create a value of CreateSolutionVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:solutionVersionArn:CreateSolutionVersionResponse', createSolutionVersionResponse_solutionVersionArn - The ARN of the new solution version.

$sel:httpStatus:CreateSolutionVersionResponse', createSolutionVersionResponse_httpStatus - The response's http status code.

Response Lenses

createSolutionVersionResponse_solutionVersionArn :: Lens' CreateSolutionVersionResponse (Maybe Text)

The ARN of the new solution version.

createSolutionVersionResponse_httpStatus :: Lens' CreateSolutionVersionResponse Int

The response's http status code.