Network.AWS.Personalize.DescribeAlgorithm

Description

Describes the given algorithm.

Creating a Request

data DescribeAlgorithm

See: newDescribeAlgorithm smart constructor.

Constructors

DescribeAlgorithm'

Fields

algorithmArn :: Text

The Amazon Resource Name (ARN) of the algorithm to describe.

newDescribeAlgorithm

Arguments

:: Text
$sel:algorithmArn:DescribeAlgorithm'

-> DescribeAlgorithm

Create a value of DescribeAlgorithm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:algorithmArn:DescribeAlgorithm', describeAlgorithm_algorithmArn - The Amazon Resource Name (ARN) of the algorithm to describe.

Request Lenses

describeAlgorithm_algorithmArn :: Lens' DescribeAlgorithm Text

The Amazon Resource Name (ARN) of the algorithm to describe.

Destructuring the Response

data DescribeAlgorithmResponse

See: newDescribeAlgorithmResponse smart constructor.

Constructors

DescribeAlgorithmResponse'

Fields

algorithm :: Maybe Algorithm

A listing of the properties of the algorithm.

httpStatus :: Int

The response's http status code.

newDescribeAlgorithmResponse

Arguments

:: Int
$sel:httpStatus:DescribeAlgorithmResponse'

-> DescribeAlgorithmResponse

Create a value of DescribeAlgorithmResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:algorithm:DescribeAlgorithmResponse', describeAlgorithmResponse_algorithm - A listing of the properties of the algorithm.

$sel:httpStatus:DescribeAlgorithmResponse', describeAlgorithmResponse_httpStatus - The response's http status code.

Response Lenses

describeAlgorithmResponse_algorithm :: Lens' DescribeAlgorithmResponse (Maybe Algorithm)

A listing of the properties of the algorithm.

describeAlgorithmResponse_httpStatus :: Lens' DescribeAlgorithmResponse Int

The response's http status code.