Network.AWS.Personalize.DescribeFeatureTransformation

Description

Describes the given feature transformation.

Creating a Request

data DescribeFeatureTransformation

See: newDescribeFeatureTransformation smart constructor.

Constructors

DescribeFeatureTransformation'

Fields

featureTransformationArn :: Text

The Amazon Resource Name (ARN) of the feature transformation to describe.

newDescribeFeatureTransformation

Arguments

:: Text
$sel:featureTransformationArn:DescribeFeatureTransformation'

-> DescribeFeatureTransformation

Create a value of DescribeFeatureTransformation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:featureTransformationArn:DescribeFeatureTransformation', describeFeatureTransformation_featureTransformationArn - The Amazon Resource Name (ARN) of the feature transformation to describe.

Request Lenses

describeFeatureTransformation_featureTransformationArn :: Lens' DescribeFeatureTransformation Text

The Amazon Resource Name (ARN) of the feature transformation to describe.

Destructuring the Response

data DescribeFeatureTransformationResponse

See: newDescribeFeatureTransformationResponse smart constructor.

Constructors

DescribeFeatureTransformationResponse'

Fields

featureTransformation :: Maybe FeatureTransformation

A listing of the FeatureTransformation properties.

httpStatus :: Int

The response's http status code.

newDescribeFeatureTransformationResponse

Arguments

:: Int
$sel:httpStatus:DescribeFeatureTransformationResponse'

-> DescribeFeatureTransformationResponse

Create a value of DescribeFeatureTransformationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:featureTransformation:DescribeFeatureTransformationResponse', describeFeatureTransformationResponse_featureTransformation - A listing of the FeatureTransformation properties.

$sel:httpStatus:DescribeFeatureTransformationResponse', describeFeatureTransformationResponse_httpStatus - The response's http status code.

Response Lenses

describeFeatureTransformationResponse_featureTransformation :: Lens' DescribeFeatureTransformationResponse (Maybe FeatureTransformation)

A listing of the FeatureTransformation properties.

describeFeatureTransformationResponse_httpStatus :: Lens' DescribeFeatureTransformationResponse Int

The response's http status code.