Network.AWS.Personalize.DescribeSolution

Description

Describes a solution. For more information on solutions, see CreateSolution.

Creating a Request

data DescribeSolution

See: newDescribeSolution smart constructor.

Constructors

DescribeSolution'

Fields

solutionArn :: Text

The Amazon Resource Name (ARN) of the solution to describe.

newDescribeSolution

Arguments

:: Text
$sel:solutionArn:DescribeSolution'

-> DescribeSolution

Create a value of DescribeSolution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:solutionArn:DescribeSolution', describeSolution_solutionArn - The Amazon Resource Name (ARN) of the solution to describe.

Request Lenses

describeSolution_solutionArn :: Lens' DescribeSolution Text

The Amazon Resource Name (ARN) of the solution to describe.

Destructuring the Response

data DescribeSolutionResponse

See: newDescribeSolutionResponse smart constructor.

Constructors

DescribeSolutionResponse'

Fields

solution :: Maybe Solution

An object that describes the solution.

httpStatus :: Int

The response's http status code.

newDescribeSolutionResponse

Arguments

:: Int
$sel:httpStatus:DescribeSolutionResponse'

-> DescribeSolutionResponse

Create a value of DescribeSolutionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:solution:DescribeSolutionResponse', describeSolutionResponse_solution - An object that describes the solution.

$sel:httpStatus:DescribeSolutionResponse', describeSolutionResponse_httpStatus - The response's http status code.

Response Lenses

describeSolutionResponse_solution :: Lens' DescribeSolutionResponse (Maybe Solution)

An object that describes the solution.

describeSolutionResponse_httpStatus :: Lens' DescribeSolutionResponse Int

The response's http status code.