Network.AWS.Personalize.ListDatasetExportJobs

Description

Returns a list of dataset export jobs that use the given dataset. When a dataset is not specified, all the dataset export jobs associated with the account are listed. The response provides the properties for each dataset export job, including the Amazon Resource Name (ARN). For more information on dataset export jobs, see CreateDatasetExportJob. For more information on datasets, see CreateDataset.

This operation returns paginated results.

Creating a Request

data ListDatasetExportJobs

See: newListDatasetExportJobs smart constructor.

Constructors

ListDatasetExportJobs'

Fields

datasetArn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset to list the dataset export jobs for.

nextToken :: Maybe Text

A token returned from the previous call to ListDatasetExportJobs for getting the next set of dataset export jobs (if they exist).

maxResults :: Maybe Natural

The maximum number of dataset export jobs to return.

newListDatasetExportJobs :: ListDatasetExportJobs

Create a value of ListDatasetExportJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetArn:ListDatasetExportJobs', listDatasetExportJobs_datasetArn - The Amazon Resource Name (ARN) of the dataset to list the dataset export jobs for.

$sel:nextToken:ListDatasetExportJobs', listDatasetExportJobs_nextToken - A token returned from the previous call to ListDatasetExportJobs for getting the next set of dataset export jobs (if they exist).

$sel:maxResults:ListDatasetExportJobs', listDatasetExportJobs_maxResults - The maximum number of dataset export jobs to return.

Request Lenses

listDatasetExportJobs_datasetArn :: Lens' ListDatasetExportJobs (Maybe Text)

The Amazon Resource Name (ARN) of the dataset to list the dataset export jobs for.

listDatasetExportJobs_nextToken :: Lens' ListDatasetExportJobs (Maybe Text)

A token returned from the previous call to ListDatasetExportJobs for getting the next set of dataset export jobs (if they exist).

listDatasetExportJobs_maxResults :: Lens' ListDatasetExportJobs (Maybe Natural)

The maximum number of dataset export jobs to return.

Destructuring the Response

data ListDatasetExportJobsResponse

See: newListDatasetExportJobsResponse smart constructor.

Constructors

ListDatasetExportJobsResponse'

Fields

nextToken :: Maybe Text

A token for getting the next set of dataset export jobs (if they exist).

datasetExportJobs :: Maybe [DatasetExportJobSummary]

The list of dataset export jobs.

httpStatus :: Int

The response's http status code.

newListDatasetExportJobsResponse

Arguments

:: Int
$sel:httpStatus:ListDatasetExportJobsResponse'

-> ListDatasetExportJobsResponse

Create a value of ListDatasetExportJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListDatasetExportJobs', listDatasetExportJobsResponse_nextToken - A token for getting the next set of dataset export jobs (if they exist).

$sel:datasetExportJobs:ListDatasetExportJobsResponse', listDatasetExportJobsResponse_datasetExportJobs - The list of dataset export jobs.

$sel:httpStatus:ListDatasetExportJobsResponse', listDatasetExportJobsResponse_httpStatus - The response's http status code.

Response Lenses

listDatasetExportJobsResponse_nextToken :: Lens' ListDatasetExportJobsResponse (Maybe Text)

A token for getting the next set of dataset export jobs (if they exist).

listDatasetExportJobsResponse_datasetExportJobs :: Lens' ListDatasetExportJobsResponse (Maybe [DatasetExportJobSummary])

The list of dataset export jobs.

listDatasetExportJobsResponse_httpStatus :: Lens' ListDatasetExportJobsResponse Int

The response's http status code.