Network.AWS.Personalize.Types.Algorithm

Description

Documentation

data Algorithm

Describes a custom algorithm.

See: newAlgorithm smart constructor.

Constructors

Algorithm'

Fields

defaultHyperParameters :: Maybe (HashMap Text Text)

Specifies the default hyperparameters.

algorithmArn :: Maybe Text

The Amazon Resource Name (ARN) of the algorithm.

trainingInputMode :: Maybe Text

The training input mode.

defaultHyperParameterRanges :: Maybe DefaultHyperParameterRanges

Specifies the default hyperparameters, their ranges, and whether they are tunable. A tunable hyperparameter can have its value determined during hyperparameter optimization (HPO).

algorithmImage :: Maybe AlgorithmImage

The URI of the Docker container for the algorithm image.

lastUpdatedDateTime :: Maybe POSIX

The date and time (in Unix time) that the algorithm was last updated.

name :: Maybe Text

The name of the algorithm.

creationDateTime :: Maybe POSIX

The date and time (in Unix time) that the algorithm was created.

defaultResourceConfig :: Maybe (HashMap Text Text)

Specifies the default maximum number of training jobs and parallel training jobs.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the role.

newAlgorithm :: Algorithm

Create a value of Algorithm with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defaultHyperParameters:Algorithm', algorithm_defaultHyperParameters - Specifies the default hyperparameters.

$sel:algorithmArn:Algorithm', algorithm_algorithmArn - The Amazon Resource Name (ARN) of the algorithm.

$sel:trainingInputMode:Algorithm', algorithm_trainingInputMode - The training input mode.

$sel:defaultHyperParameterRanges:Algorithm', algorithm_defaultHyperParameterRanges - Specifies the default hyperparameters, their ranges, and whether they are tunable. A tunable hyperparameter can have its value determined during hyperparameter optimization (HPO).

$sel:algorithmImage:Algorithm', algorithm_algorithmImage - The URI of the Docker container for the algorithm image.

$sel:lastUpdatedDateTime:Algorithm', algorithm_lastUpdatedDateTime - The date and time (in Unix time) that the algorithm was last updated.

$sel:name:Algorithm', algorithm_name - The name of the algorithm.

$sel:creationDateTime:Algorithm', algorithm_creationDateTime - The date and time (in Unix time) that the algorithm was created.

$sel:defaultResourceConfig:Algorithm', algorithm_defaultResourceConfig - Specifies the default maximum number of training jobs and parallel training jobs.

$sel:roleArn:Algorithm', algorithm_roleArn - The Amazon Resource Name (ARN) of the role.

algorithm_defaultHyperParameters :: Lens' Algorithm (Maybe (HashMap Text Text))

Specifies the default hyperparameters.

algorithm_algorithmArn :: Lens' Algorithm (Maybe Text)

The Amazon Resource Name (ARN) of the algorithm.

algorithm_trainingInputMode :: Lens' Algorithm (Maybe Text)

The training input mode.

algorithm_defaultHyperParameterRanges :: Lens' Algorithm (Maybe DefaultHyperParameterRanges)

Specifies the default hyperparameters, their ranges, and whether they are tunable. A tunable hyperparameter can have its value determined during hyperparameter optimization (HPO).

algorithm_algorithmImage :: Lens' Algorithm (Maybe AlgorithmImage)

The URI of the Docker container for the algorithm image.

algorithm_lastUpdatedDateTime :: Lens' Algorithm (Maybe UTCTime)

The date and time (in Unix time) that the algorithm was last updated.

algorithm_name :: Lens' Algorithm (Maybe Text)

The name of the algorithm.

algorithm_creationDateTime :: Lens' Algorithm (Maybe UTCTime)

The date and time (in Unix time) that the algorithm was created.

algorithm_defaultResourceConfig :: Lens' Algorithm (Maybe (HashMap Text Text))

Specifies the default maximum number of training jobs and parallel training jobs.

algorithm_roleArn :: Lens' Algorithm (Maybe Text)

The Amazon Resource Name (ARN) of the role.