Network.AWS.Personalize.Types.BatchInferenceJobOutput

Description

Documentation

data BatchInferenceJobOutput

The output configuration parameters of a batch inference job.

See: newBatchInferenceJobOutput smart constructor.

Constructors

BatchInferenceJobOutput'

Fields

s3DataDestination :: S3DataConfig

Information on the Amazon S3 bucket in which the batch inference job's output is stored.

newBatchInferenceJobOutput

Arguments

:: S3DataConfig
$sel:s3DataDestination:BatchInferenceJobOutput'

-> BatchInferenceJobOutput

Create a value of BatchInferenceJobOutput with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:s3DataDestination:BatchInferenceJobOutput', batchInferenceJobOutput_s3DataDestination - Information on the Amazon S3 bucket in which the batch inference job's output is stored.

batchInferenceJobOutput_s3DataDestination :: Lens' BatchInferenceJobOutput S3DataConfig

Information on the Amazon S3 bucket in which the batch inference job's output is stored.