Network.AWS.Personalize.Types.CampaignUpdateSummary

Description

Documentation

data CampaignUpdateSummary

Provides a summary of the properties of a campaign update. For a complete listing, call the DescribeCampaign API.

See: newCampaignUpdateSummary smart constructor.

Constructors

CampaignUpdateSummary'

Fields

failureReason :: Maybe Text

If a campaign update fails, the reason behind the failure.

status :: Maybe Text

The status of the campaign update.

A campaign update can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

DELETE PENDING > DELETE IN_PROGRESS

lastUpdatedDateTime :: Maybe POSIX

The date and time (in Unix time) that the campaign update was last updated.

campaignConfig :: Maybe CampaignConfig

minProvisionedTPS :: Maybe Natural

Specifies the requested minimum provisioned transactions (recommendations) per second that Amazon Personalize will support.

creationDateTime :: Maybe POSIX

The date and time (in Unix time) that the campaign update was created.

solutionVersionArn :: Maybe Text

The Amazon Resource Name (ARN) of the deployed solution version.

newCampaignUpdateSummary :: CampaignUpdateSummary

Create a value of CampaignUpdateSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:CampaignUpdateSummary', campaignUpdateSummary_failureReason - If a campaign update fails, the reason behind the failure.

$sel:status:CampaignUpdateSummary', campaignUpdateSummary_status - The status of the campaign update.

A campaign update can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

DELETE PENDING > DELETE IN_PROGRESS

$sel:lastUpdatedDateTime:CampaignUpdateSummary', campaignUpdateSummary_lastUpdatedDateTime - The date and time (in Unix time) that the campaign update was last updated.

$sel:campaignConfig:CampaignUpdateSummary', campaignUpdateSummary_campaignConfig - Undocumented member.

$sel:minProvisionedTPS:CampaignUpdateSummary', campaignUpdateSummary_minProvisionedTPS - Specifies the requested minimum provisioned transactions (recommendations) per second that Amazon Personalize will support.

$sel:creationDateTime:CampaignUpdateSummary', campaignUpdateSummary_creationDateTime - The date and time (in Unix time) that the campaign update was created.

$sel:solutionVersionArn:CampaignUpdateSummary', campaignUpdateSummary_solutionVersionArn - The Amazon Resource Name (ARN) of the deployed solution version.

campaignUpdateSummary_failureReason :: Lens' CampaignUpdateSummary (Maybe Text)

If a campaign update fails, the reason behind the failure.

campaignUpdateSummary_status :: Lens' CampaignUpdateSummary (Maybe Text)

The status of the campaign update.

A campaign update can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

DELETE PENDING > DELETE IN_PROGRESS

campaignUpdateSummary_lastUpdatedDateTime :: Lens' CampaignUpdateSummary (Maybe UTCTime)

The date and time (in Unix time) that the campaign update was last updated.

campaignUpdateSummary_campaignConfig :: Lens' CampaignUpdateSummary (Maybe CampaignConfig)

Undocumented member.

campaignUpdateSummary_minProvisionedTPS :: Lens' CampaignUpdateSummary (Maybe Natural)

Specifies the requested minimum provisioned transactions (recommendations) per second that Amazon Personalize will support.

campaignUpdateSummary_creationDateTime :: Lens' CampaignUpdateSummary (Maybe UTCTime)

The date and time (in Unix time) that the campaign update was created.

campaignUpdateSummary_solutionVersionArn :: Lens' CampaignUpdateSummary (Maybe Text)

The Amazon Resource Name (ARN) of the deployed solution version.