Network.AWS.Personalize.Types.DatasetExportJobSummary

Description

Documentation

data DatasetExportJobSummary

Provides a summary of the properties of a dataset export job. For a complete listing, call the DescribeDatasetExportJob API.

See: newDatasetExportJobSummary smart constructor.

Constructors

DatasetExportJobSummary'

Fields

failureReason :: Maybe Text

If a dataset export job fails, the reason behind the failure.

status :: Maybe Text

The status of the dataset export job.

A dataset export job can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

datasetExportJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset export job.

jobName :: Maybe Text

The name of the dataset export job.

lastUpdatedDateTime :: Maybe POSIX

The date and time (in Unix time) that the dataset export job status was last updated.

creationDateTime :: Maybe POSIX

The date and time (in Unix time) that the dataset export job was created.

newDatasetExportJobSummary :: DatasetExportJobSummary

Create a value of DatasetExportJobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:DatasetExportJobSummary', datasetExportJobSummary_failureReason - If a dataset export job fails, the reason behind the failure.

$sel:status:DatasetExportJobSummary', datasetExportJobSummary_status - The status of the dataset export job.

A dataset export job can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

$sel:datasetExportJobArn:DatasetExportJobSummary', datasetExportJobSummary_datasetExportJobArn - The Amazon Resource Name (ARN) of the dataset export job.

$sel:jobName:DatasetExportJobSummary', datasetExportJobSummary_jobName - The name of the dataset export job.

$sel:lastUpdatedDateTime:DatasetExportJobSummary', datasetExportJobSummary_lastUpdatedDateTime - The date and time (in Unix time) that the dataset export job status was last updated.

$sel:creationDateTime:DatasetExportJobSummary', datasetExportJobSummary_creationDateTime - The date and time (in Unix time) that the dataset export job was created.

datasetExportJobSummary_failureReason :: Lens' DatasetExportJobSummary (Maybe Text)

If a dataset export job fails, the reason behind the failure.

datasetExportJobSummary_status :: Lens' DatasetExportJobSummary (Maybe Text)

The status of the dataset export job.

A dataset export job can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

datasetExportJobSummary_datasetExportJobArn :: Lens' DatasetExportJobSummary (Maybe Text)

The Amazon Resource Name (ARN) of the dataset export job.

datasetExportJobSummary_jobName :: Lens' DatasetExportJobSummary (Maybe Text)

The name of the dataset export job.

datasetExportJobSummary_lastUpdatedDateTime :: Lens' DatasetExportJobSummary (Maybe UTCTime)

The date and time (in Unix time) that the dataset export job status was last updated.

datasetExportJobSummary_creationDateTime :: Lens' DatasetExportJobSummary (Maybe UTCTime)

The date and time (in Unix time) that the dataset export job was created.