Network.AWS.Personalize.Types.DatasetGroup

Description

Documentation

data DatasetGroup

A dataset group is a collection of related datasets (Interactions, User, and Item). You create a dataset group by calling CreateDatasetGroup. You then create a dataset and add it to a dataset group by calling CreateDataset. The dataset group is used to create and train a solution by calling CreateSolution. A dataset group can contain only one of each type of dataset.

You can specify an Key Management Service (KMS) key to encrypt the datasets in the group.

See: newDatasetGroup smart constructor.

Constructors

DatasetGroup'

Fields

failureReason :: Maybe Text

If creating a dataset group fails, provides the reason why.

status :: Maybe Text

The current status of the dataset group.

A dataset group can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

DELETE PENDING

kmsKeyArn :: Maybe Text

The Amazon Resource Name (ARN) of the Key Management Service (KMS) key used to encrypt the datasets.

lastUpdatedDateTime :: Maybe POSIX

The last update date and time (in Unix time) of the dataset group.

name :: Maybe Text

The name of the dataset group.

creationDateTime :: Maybe POSIX

The creation date and time (in Unix time) of the dataset group.

datasetGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset group.

roleArn :: Maybe Text

The ARN of the IAM role that has permissions to create the dataset group.

newDatasetGroup :: DatasetGroup

Create a value of DatasetGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:DatasetGroup', datasetGroup_failureReason - If creating a dataset group fails, provides the reason why.

$sel:status:DatasetGroup', datasetGroup_status - The current status of the dataset group.

A dataset group can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

DELETE PENDING

$sel:kmsKeyArn:DatasetGroup', datasetGroup_kmsKeyArn - The Amazon Resource Name (ARN) of the Key Management Service (KMS) key used to encrypt the datasets.

$sel:lastUpdatedDateTime:DatasetGroup', datasetGroup_lastUpdatedDateTime - The last update date and time (in Unix time) of the dataset group.

$sel:name:DatasetGroup', datasetGroup_name - The name of the dataset group.

$sel:creationDateTime:DatasetGroup', datasetGroup_creationDateTime - The creation date and time (in Unix time) of the dataset group.

$sel:datasetGroupArn:DatasetGroup', datasetGroup_datasetGroupArn - The Amazon Resource Name (ARN) of the dataset group.

$sel:roleArn:DatasetGroup', datasetGroup_roleArn - The ARN of the IAM role that has permissions to create the dataset group.

datasetGroup_failureReason :: Lens' DatasetGroup (Maybe Text)

If creating a dataset group fails, provides the reason why.

datasetGroup_status :: Lens' DatasetGroup (Maybe Text)

The current status of the dataset group.

A dataset group can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

DELETE PENDING

datasetGroup_kmsKeyArn :: Lens' DatasetGroup (Maybe Text)

The Amazon Resource Name (ARN) of the Key Management Service (KMS) key used to encrypt the datasets.

datasetGroup_lastUpdatedDateTime :: Lens' DatasetGroup (Maybe UTCTime)

The last update date and time (in Unix time) of the dataset group.

datasetGroup_name :: Lens' DatasetGroup (Maybe Text)

The name of the dataset group.

datasetGroup_creationDateTime :: Lens' DatasetGroup (Maybe UTCTime)

The creation date and time (in Unix time) of the dataset group.

datasetGroup_datasetGroupArn :: Lens' DatasetGroup (Maybe Text)

The Amazon Resource Name (ARN) of the dataset group.

datasetGroup_roleArn :: Lens' DatasetGroup (Maybe Text)

The ARN of the IAM role that has permissions to create the dataset group.