Network.AWS.Personalize.Types.DatasetImportJob

Description

Documentation

data DatasetImportJob

Describes a job that imports training data from a data source (Amazon S3 bucket) to an Amazon Personalize dataset. For more information, see CreateDatasetImportJob.

A dataset import job can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

See: newDatasetImportJob smart constructor.

Constructors

DatasetImportJob'

Fields

failureReason :: Maybe Text

If a dataset import job fails, provides the reason why.

status :: Maybe Text

The status of the dataset import job.

A dataset import job can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

datasetArn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset that receives the imported data.

jobName :: Maybe Text

The name of the import job.

lastUpdatedDateTime :: Maybe POSIX

The date and time (in Unix time) the dataset was last updated.

datasetImportJobArn :: Maybe Text

The ARN of the dataset import job.

dataSource :: Maybe DataSource

The Amazon S3 bucket that contains the training data to import.

creationDateTime :: Maybe POSIX

The creation date and time (in Unix time) of the dataset import job.

roleArn :: Maybe Text

The ARN of the IAM role that has permissions to read from the Amazon S3 data source.

newDatasetImportJob :: DatasetImportJob

Create a value of DatasetImportJob with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:DatasetImportJob', datasetImportJob_failureReason - If a dataset import job fails, provides the reason why.

$sel:status:DatasetImportJob', datasetImportJob_status - The status of the dataset import job.

A dataset import job can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

$sel:datasetArn:DatasetImportJob', datasetImportJob_datasetArn - The Amazon Resource Name (ARN) of the dataset that receives the imported data.

$sel:jobName:DatasetImportJob', datasetImportJob_jobName - The name of the import job.

$sel:lastUpdatedDateTime:DatasetImportJob', datasetImportJob_lastUpdatedDateTime - The date and time (in Unix time) the dataset was last updated.

$sel:datasetImportJobArn:DatasetImportJob', datasetImportJob_datasetImportJobArn - The ARN of the dataset import job.

$sel:dataSource:DatasetImportJob', datasetImportJob_dataSource - The Amazon S3 bucket that contains the training data to import.

$sel:creationDateTime:DatasetImportJob', datasetImportJob_creationDateTime - The creation date and time (in Unix time) of the dataset import job.

$sel:roleArn:DatasetImportJob', datasetImportJob_roleArn - The ARN of the IAM role that has permissions to read from the Amazon S3 data source.

datasetImportJob_failureReason :: Lens' DatasetImportJob (Maybe Text)

If a dataset import job fails, provides the reason why.

datasetImportJob_status :: Lens' DatasetImportJob (Maybe Text)

The status of the dataset import job.

A dataset import job can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

datasetImportJob_datasetArn :: Lens' DatasetImportJob (Maybe Text)

The Amazon Resource Name (ARN) of the dataset that receives the imported data.

datasetImportJob_jobName :: Lens' DatasetImportJob (Maybe Text)

The name of the import job.

datasetImportJob_lastUpdatedDateTime :: Lens' DatasetImportJob (Maybe UTCTime)

The date and time (in Unix time) the dataset was last updated.

datasetImportJob_datasetImportJobArn :: Lens' DatasetImportJob (Maybe Text)

The ARN of the dataset import job.

datasetImportJob_dataSource :: Lens' DatasetImportJob (Maybe DataSource)

The Amazon S3 bucket that contains the training data to import.

datasetImportJob_creationDateTime :: Lens' DatasetImportJob (Maybe UTCTime)

The creation date and time (in Unix time) of the dataset import job.

datasetImportJob_roleArn :: Lens' DatasetImportJob (Maybe Text)

The ARN of the IAM role that has permissions to read from the Amazon S3 data source.