Network.AWS.Personalize.Types.DatasetImportJobSummary

Description

Documentation

data DatasetImportJobSummary

Provides a summary of the properties of a dataset import job. For a complete listing, call the DescribeDatasetImportJob API.

See: newDatasetImportJobSummary smart constructor.

Constructors

DatasetImportJobSummary'

Fields

failureReason :: Maybe Text

If a dataset import job fails, the reason behind the failure.

status :: Maybe Text

The status of the dataset import job.

A dataset import job can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

jobName :: Maybe Text

The name of the dataset import job.

lastUpdatedDateTime :: Maybe POSIX

The date and time (in Unix time) that the dataset import job status was last updated.

datasetImportJobArn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset import job.

creationDateTime :: Maybe POSIX

The date and time (in Unix time) that the dataset import job was created.

newDatasetImportJobSummary :: DatasetImportJobSummary

Create a value of DatasetImportJobSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:failureReason:DatasetImportJobSummary', datasetImportJobSummary_failureReason - If a dataset import job fails, the reason behind the failure.

$sel:status:DatasetImportJobSummary', datasetImportJobSummary_status - The status of the dataset import job.

A dataset import job can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

$sel:jobName:DatasetImportJobSummary', datasetImportJobSummary_jobName - The name of the dataset import job.

$sel:lastUpdatedDateTime:DatasetImportJobSummary', datasetImportJobSummary_lastUpdatedDateTime - The date and time (in Unix time) that the dataset import job status was last updated.

$sel:datasetImportJobArn:DatasetImportJobSummary', datasetImportJobSummary_datasetImportJobArn - The Amazon Resource Name (ARN) of the dataset import job.

$sel:creationDateTime:DatasetImportJobSummary', datasetImportJobSummary_creationDateTime - The date and time (in Unix time) that the dataset import job was created.

datasetImportJobSummary_failureReason :: Lens' DatasetImportJobSummary (Maybe Text)

If a dataset import job fails, the reason behind the failure.

datasetImportJobSummary_status :: Lens' DatasetImportJobSummary (Maybe Text)

The status of the dataset import job.

A dataset import job can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

datasetImportJobSummary_jobName :: Lens' DatasetImportJobSummary (Maybe Text)

The name of the dataset import job.

datasetImportJobSummary_lastUpdatedDateTime :: Lens' DatasetImportJobSummary (Maybe UTCTime)

The date and time (in Unix time) that the dataset import job status was last updated.

datasetImportJobSummary_datasetImportJobArn :: Lens' DatasetImportJobSummary (Maybe Text)

The Amazon Resource Name (ARN) of the dataset import job.

datasetImportJobSummary_creationDateTime :: Lens' DatasetImportJobSummary (Maybe UTCTime)

The date and time (in Unix time) that the dataset import job was created.