Network.AWS.Personalize.Types.Recipe

Description

Documentation

data Recipe

Provides information about a recipe. Each recipe provides an algorithm that Amazon Personalize uses in model training when you use the CreateSolution operation.

See: newRecipe smart constructor.

Constructors

Recipe'

Fields

status :: Maybe Text

The status of the recipe.

algorithmArn :: Maybe Text

The Amazon Resource Name (ARN) of the algorithm that Amazon Personalize uses to train the model.

recipeArn :: Maybe Text

The Amazon Resource Name (ARN) of the recipe.

featureTransformationArn :: Maybe Text

The ARN of the FeatureTransformation object.

lastUpdatedDateTime :: Maybe POSIX

The date and time (in Unix format) that the recipe was last updated.

name :: Maybe Text

The name of the recipe.

creationDateTime :: Maybe POSIX

The date and time (in Unix format) that the recipe was created.

recipeType :: Maybe Text

One of the following values:

PERSONALIZED_RANKING

RELATED_ITEMS

USER_PERSONALIZATION

description :: Maybe Text

The description of the recipe.

newRecipe :: Recipe

Create a value of Recipe with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Recipe', recipe_status - The status of the recipe.

$sel:algorithmArn:Recipe', recipe_algorithmArn - The Amazon Resource Name (ARN) of the algorithm that Amazon Personalize uses to train the model.

$sel:recipeArn:Recipe', recipe_recipeArn - The Amazon Resource Name (ARN) of the recipe.

$sel:featureTransformationArn:Recipe', recipe_featureTransformationArn - The ARN of the FeatureTransformation object.

$sel:lastUpdatedDateTime:Recipe', recipe_lastUpdatedDateTime - The date and time (in Unix format) that the recipe was last updated.

$sel:name:Recipe', recipe_name - The name of the recipe.

$sel:creationDateTime:Recipe', recipe_creationDateTime - The date and time (in Unix format) that the recipe was created.

$sel:recipeType:Recipe', recipe_recipeType - One of the following values:

PERSONALIZED_RANKING

RELATED_ITEMS

USER_PERSONALIZATION

$sel:description:Recipe', recipe_description - The description of the recipe.

recipe_status :: Lens' Recipe (Maybe Text)

The status of the recipe.

recipe_algorithmArn :: Lens' Recipe (Maybe Text)

The Amazon Resource Name (ARN) of the algorithm that Amazon Personalize uses to train the model.

recipe_recipeArn :: Lens' Recipe (Maybe Text)

The Amazon Resource Name (ARN) of the recipe.

recipe_featureTransformationArn :: Lens' Recipe (Maybe Text)

The ARN of the FeatureTransformation object.

recipe_lastUpdatedDateTime :: Lens' Recipe (Maybe UTCTime)

The date and time (in Unix format) that the recipe was last updated.

recipe_name :: Lens' Recipe (Maybe Text)

The name of the recipe.

recipe_creationDateTime :: Lens' Recipe (Maybe UTCTime)

The date and time (in Unix format) that the recipe was created.

recipe_recipeType :: Lens' Recipe (Maybe Text)

One of the following values:

PERSONALIZED_RANKING

RELATED_ITEMS

USER_PERSONALIZATION

recipe_description :: Lens' Recipe (Maybe Text)

The description of the recipe.