Network.AWS.Personalize.Types.Solution

Description

Documentation

data Solution

An object that provides information about a solution. A solution is a trained model that can be deployed as a campaign.

See: newSolution smart constructor.

Constructors

Solution'

Fields

solutionArn :: Maybe Text

The ARN of the solution.

status :: Maybe Text

The status of the solution.

A solution can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

DELETE PENDING > DELETE IN_PROGRESS

performAutoML :: Maybe Bool

When true, Amazon Personalize performs a search for the best USER_PERSONALIZATION recipe from the list specified in the solution configuration (recipeArn must not be specified). When false (the default), Amazon Personalize uses recipeArn for training.

recipeArn :: Maybe Text

The ARN of the recipe used to create the solution.

lastUpdatedDateTime :: Maybe POSIX

The date and time (in Unix time) that the solution was last updated.

eventType :: Maybe Text

The event type (for example, 'click' or 'like') that is used for training the model. If no eventType is provided, Amazon Personalize uses all interactions for training with equal weight regardless of type.

name :: Maybe Text

The name of the solution.

autoMLResult :: Maybe AutoMLResult

When performAutoML is true, specifies the best recipe found.

creationDateTime :: Maybe POSIX

The creation date and time (in Unix time) of the solution.

datasetGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset group that provides the training data.

latestSolutionVersion :: Maybe SolutionVersionSummary

Describes the latest version of the solution, including the status and the ARN.

solutionConfig :: Maybe SolutionConfig

Describes the configuration properties for the solution.

performHPO :: Maybe Bool

Whether to perform hyperparameter optimization (HPO) on the chosen recipe. The default is false.

newSolution :: Solution

Create a value of Solution with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:solutionArn:Solution', solution_solutionArn - The ARN of the solution.

$sel:status:Solution', solution_status - The status of the solution.

A solution can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

DELETE PENDING > DELETE IN_PROGRESS

$sel:performAutoML:Solution', solution_performAutoML - When true, Amazon Personalize performs a search for the best USER_PERSONALIZATION recipe from the list specified in the solution configuration (recipeArn must not be specified). When false (the default), Amazon Personalize uses recipeArn for training.

$sel:recipeArn:Solution', solution_recipeArn - The ARN of the recipe used to create the solution.

$sel:lastUpdatedDateTime:Solution', solution_lastUpdatedDateTime - The date and time (in Unix time) that the solution was last updated.

$sel:eventType:Solution', solution_eventType - The event type (for example, 'click' or 'like') that is used for training the model. If no eventType is provided, Amazon Personalize uses all interactions for training with equal weight regardless of type.

$sel:name:Solution', solution_name - The name of the solution.

$sel:autoMLResult:Solution', solution_autoMLResult - When performAutoML is true, specifies the best recipe found.

$sel:creationDateTime:Solution', solution_creationDateTime - The creation date and time (in Unix time) of the solution.

$sel:datasetGroupArn:Solution', solution_datasetGroupArn - The Amazon Resource Name (ARN) of the dataset group that provides the training data.

$sel:latestSolutionVersion:Solution', solution_latestSolutionVersion - Describes the latest version of the solution, including the status and the ARN.

$sel:solutionConfig:Solution', solution_solutionConfig - Describes the configuration properties for the solution.

$sel:performHPO:Solution', solution_performHPO - Whether to perform hyperparameter optimization (HPO) on the chosen recipe. The default is false.

solution_solutionArn :: Lens' Solution (Maybe Text)

The ARN of the solution.

solution_status :: Lens' Solution (Maybe Text)

The status of the solution.

A solution can be in one of the following states:

CREATE PENDING > CREATE IN_PROGRESS > ACTIVE -or- CREATE FAILED

DELETE PENDING > DELETE IN_PROGRESS

solution_performAutoML :: Lens' Solution (Maybe Bool)

When true, Amazon Personalize performs a search for the best USER_PERSONALIZATION recipe from the list specified in the solution configuration (recipeArn must not be specified). When false (the default), Amazon Personalize uses recipeArn for training.

solution_recipeArn :: Lens' Solution (Maybe Text)

The ARN of the recipe used to create the solution.

solution_lastUpdatedDateTime :: Lens' Solution (Maybe UTCTime)

The date and time (in Unix time) that the solution was last updated.

solution_eventType :: Lens' Solution (Maybe Text)

The event type (for example, 'click' or 'like') that is used for training the model. If no eventType is provided, Amazon Personalize uses all interactions for training with equal weight regardless of type.

solution_name :: Lens' Solution (Maybe Text)

The name of the solution.

solution_autoMLResult :: Lens' Solution (Maybe AutoMLResult)

When performAutoML is true, specifies the best recipe found.

solution_creationDateTime :: Lens' Solution (Maybe UTCTime)

The creation date and time (in Unix time) of the solution.

solution_datasetGroupArn :: Lens' Solution (Maybe Text)

The Amazon Resource Name (ARN) of the dataset group that provides the training data.

solution_latestSolutionVersion :: Lens' Solution (Maybe SolutionVersionSummary)

Describes the latest version of the solution, including the status and the ARN.

solution_solutionConfig :: Lens' Solution (Maybe SolutionConfig)

Describes the configuration properties for the solution.

solution_performHPO :: Lens' Solution (Maybe Bool)

Whether to perform hyperparameter optimization (HPO) on the chosen recipe. The default is false.