Network.AWS.Personalize.Types.SolutionConfig

Description

Documentation

data SolutionConfig

Describes the configuration properties for the solution.

See: newSolutionConfig smart constructor.

Constructors

SolutionConfig'

Fields

featureTransformationParameters :: Maybe (HashMap Text Text)

Lists the feature transformation parameters.

hpoConfig :: Maybe HPOConfig

Describes the properties for hyperparameter optimization (HPO).

eventValueThreshold :: Maybe Text

Only events with a value greater than or equal to this threshold are used for training a model.

autoMLConfig :: Maybe AutoMLConfig

The AutoMLConfig object containing a list of recipes to search when AutoML is performed.

algorithmHyperParameters :: Maybe (HashMap Text Text)

Lists the hyperparameter names and ranges.

optimizationObjective :: Maybe OptimizationObjective

Describes the additional objective for the solution, such as maximizing streaming minutes or increasing revenue. For more information see Optimizing a solution.

newSolutionConfig :: SolutionConfig

Create a value of SolutionConfig with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:featureTransformationParameters:SolutionConfig', solutionConfig_featureTransformationParameters - Lists the feature transformation parameters.

$sel:hpoConfig:SolutionConfig', solutionConfig_hpoConfig - Describes the properties for hyperparameter optimization (HPO).

$sel:eventValueThreshold:SolutionConfig', solutionConfig_eventValueThreshold - Only events with a value greater than or equal to this threshold are used for training a model.

$sel:autoMLConfig:SolutionConfig', solutionConfig_autoMLConfig - The AutoMLConfig object containing a list of recipes to search when AutoML is performed.

$sel:algorithmHyperParameters:SolutionConfig', solutionConfig_algorithmHyperParameters - Lists the hyperparameter names and ranges.

$sel:optimizationObjective:SolutionConfig', solutionConfig_optimizationObjective - Describes the additional objective for the solution, such as maximizing streaming minutes or increasing revenue. For more information see Optimizing a solution.

solutionConfig_featureTransformationParameters :: Lens' SolutionConfig (Maybe (HashMap Text Text))

Lists the feature transformation parameters.

solutionConfig_hpoConfig :: Lens' SolutionConfig (Maybe HPOConfig)

Describes the properties for hyperparameter optimization (HPO).

solutionConfig_eventValueThreshold :: Lens' SolutionConfig (Maybe Text)

Only events with a value greater than or equal to this threshold are used for training a model.

solutionConfig_autoMLConfig :: Lens' SolutionConfig (Maybe AutoMLConfig)

The AutoMLConfig object containing a list of recipes to search when AutoML is performed.

solutionConfig_algorithmHyperParameters :: Lens' SolutionConfig (Maybe (HashMap Text Text))

Lists the hyperparameter names and ranges.

solutionConfig_optimizationObjective :: Lens' SolutionConfig (Maybe OptimizationObjective)

Describes the additional objective for the solution, such as maximizing streaming minutes or increasing revenue. For more information see Optimizing a solution.