Network.AWS.PersonalizeEvents.Lens

Description

Operations

PutUsers

putUsers_datasetArn :: Lens' PutUsers Text

The Amazon Resource Name (ARN) of the Users dataset you are adding the user or users to.

putUsers_users :: Lens' PutUsers (NonEmpty User)

A list of user data.

PutItems

putItems_datasetArn :: Lens' PutItems Text

The Amazon Resource Name (ARN) of the Items dataset you are adding the item or items to.

putItems_items :: Lens' PutItems (NonEmpty Item)

A list of item data.

PutEvents

putEvents_userId :: Lens' PutEvents (Maybe Text)

The user associated with the event.

putEvents_trackingId :: Lens' PutEvents Text

The tracking ID for the event. The ID is generated by a call to the CreateEventTracker API.

putEvents_sessionId :: Lens' PutEvents Text

The session ID associated with the user's visit. Your application generates the sessionId when a user first visits your website or uses your application. Amazon Personalize uses the sessionId to associate events with the user before they log in. For more information, see Recording Events.

putEvents_eventList :: Lens' PutEvents (NonEmpty Event)

A list of event data from the session.

Types

Event

event_recommendationId :: Lens' Event (Maybe Text)

The ID of the recommendation.

event_eventValue :: Lens' Event (Maybe Double)

The event value that corresponds to the EVENT_VALUE field of the Interactions schema.

event_itemId :: Lens' Event (Maybe Text)

The item ID key that corresponds to the ITEM_ID field of the Interactions schema.

event_impression :: Lens' Event (Maybe (NonEmpty Text))

A list of item IDs that represents the sequence of items you have shown the user. For example, ["itemId1", "itemId2", "itemId3"].

event_eventId :: Lens' Event (Maybe Text)

An ID associated with the event. If an event ID is not provided, Amazon Personalize generates a unique ID for the event. An event ID is not used as an input to the model. Amazon Personalize uses the event ID to distinquish unique events. Any subsequent events after the first with the same event ID are not used in model training.

event_properties :: Lens' Event (Maybe Text)

A string map of event-specific data that you might choose to record. For example, if a user rates a movie on your site, other than movie ID (itemId) and rating (eventValue) , you might also send the number of movie ratings made by the user.

Each item in the map consists of a key-value pair. For example,

{"numberOfRatings": "12"}

The keys use camel case names that match the fields in the Interactions schema. In the above example, the numberOfRatings would match the 'NUMBER_OF_RATINGS' field defined in the Interactions schema.

event_eventType :: Lens' Event Text

The type of event, such as click or download. This property corresponds to the EVENT_TYPE field of your Interactions schema and depends on the types of events you are tracking.

event_sentAt :: Lens' Event UTCTime

The timestamp (in Unix time) on the client side when the event occurred.

Item

item_properties :: Lens' Item (Maybe Text)

A string map of item-specific metadata. Each element in the map consists of a key-value pair. For example, {"numberOfRatings": "12"}.

The keys use camel case names that match the fields in the schema for the Items dataset. In the previous example, the numberOfRatings matches the 'NUMBER_OF_RATINGS' field defined in the Items schema. For categorical string data, to include multiple categories for a single item, separate each category with a pipe separator (|). For example, \"Horror|Action\".

item_itemId :: Lens' Item Text

The ID associated with the item.

User

user_properties :: Lens' User (Maybe Text)

A string map of user-specific metadata. Each element in the map consists of a key-value pair. For example, {"numberOfVideosWatched": "45"}.

The keys use camel case names that match the fields in the schema for the Users dataset. In the previous example, the numberOfVideosWatched matches the 'NUMBER_OF_VIDEOS_WATCHED' field defined in the Users schema. For categorical string data, to include multiple categories for a single user, separate each category with a pipe separator (|). For example, \"Member|Frequent shopper\".

user_userId :: Lens' User Text

The ID associated with the user.