Network.AWS.PersonalizeEvents.PutItems

Description

Adds one or more items to an Items dataset. For more information see Importing Items Incrementally.

Creating a Request

data PutItems

See: newPutItems smart constructor.

Constructors

PutItems'

Fields

datasetArn :: Text

The Amazon Resource Name (ARN) of the Items dataset you are adding the item or items to.

items :: NonEmpty Item

A list of item data.

newPutItems

Arguments

:: Text
$sel:datasetArn:PutItems'

-> NonEmpty Item
$sel:items:PutItems'

-> PutItems

Create a value of PutItems with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:datasetArn:PutItems', putItems_datasetArn - The Amazon Resource Name (ARN) of the Items dataset you are adding the item or items to.

$sel:items:PutItems', putItems_items - A list of item data.

Request Lenses

putItems_datasetArn :: Lens' PutItems Text

The Amazon Resource Name (ARN) of the Items dataset you are adding the item or items to.

putItems_items :: Lens' PutItems (NonEmpty Item)

A list of item data.

Destructuring the Response

data PutItemsResponse

See: newPutItemsResponse smart constructor.

Constructors

PutItemsResponse'

newPutItemsResponse :: PutItemsResponse

Create a value of PutItemsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.