Network.AWS.PersonalizeRuntime

Description

Derived from API version 2018-05-22 of the AWS service descriptions, licensed under Apache 2.0.

Service Configuration

defaultService :: Service

API version 2018-05-22 of the Amazon Personalize Runtime SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by PersonalizeRuntime.

InvalidInputException

_InvalidInputException :: AsError a => Getting (First ServiceError) a ServiceError

Provide a valid value for the field or parameter.

ResourceNotFoundException

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource does not exist.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

GetRecommendations

data GetRecommendations

See: newGetRecommendations smart constructor.

Constructors

GetRecommendations' (Maybe (HashMap Text (Sensitive Text))) (Maybe Text) (Maybe Text) (Maybe Natural) (Maybe Text) (Maybe (HashMap Text (Sensitive Text))) Text

newGetRecommendations

Arguments

:: Text
$sel:campaignArn:GetRecommendations'

-> GetRecommendations

Create a value of GetRecommendations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:context:GetRecommendations', getRecommendations_context - The contextual metadata to use when getting recommendations. Contextual metadata includes any interaction information that might be relevant when getting a user's recommendations, such as the user's current location or device type.

$sel:itemId:GetRecommendations', getRecommendations_itemId - The item ID to provide recommendations for.

Required for RELATED_ITEMS recipe type.

$sel:userId:GetRecommendations', getRecommendations_userId - The user ID to provide recommendations for.

Required for USER_PERSONALIZATION recipe type.

$sel:numResults:GetRecommendations', getRecommendations_numResults - The number of results to return. The default is 25. The maximum is 500.

$sel:filterArn:GetRecommendations', getRecommendations_filterArn - The ARN of the filter to apply to the returned recommendations. For more information, see Filtering Recommendations.

When using this parameter, be sure the filter resource is ACTIVE.

$sel:filterValues:GetRecommendations', getRecommendations_filterValues - The values to use when filtering recommendations. For each placeholder parameter in your filter expression, provide the parameter name (in matching case) as a key and the filter value(s) as the corresponding value. Separate multiple values for one parameter with a comma.

For filter expressions that use an INCLUDE element to include items, you must provide values for all parameters that are defined in the expression. For filters with expressions that use an EXCLUDE element to exclude items, you can omit the filter-values.In this case, Amazon Personalize doesn't use that portion of the expression to filter recommendations.

For more information, see Filtering Recommendations.

$sel:campaignArn:GetRecommendations', getRecommendations_campaignArn - The Amazon Resource Name (ARN) of the campaign to use for getting recommendations.

data GetRecommendationsResponse

See: newGetRecommendationsResponse smart constructor.

Constructors

GetRecommendationsResponse' (Maybe Text) (Maybe [PredictedItem]) Int

newGetRecommendationsResponse

Arguments

:: Int
$sel:httpStatus:GetRecommendationsResponse'

-> GetRecommendationsResponse

Create a value of GetRecommendationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recommendationId:GetRecommendationsResponse', getRecommendationsResponse_recommendationId - The ID of the recommendation.

$sel:itemList:GetRecommendationsResponse', getRecommendationsResponse_itemList - A list of recommendations sorted in ascending order by prediction score. There can be a maximum of 500 items in the list.

$sel:httpStatus:GetRecommendationsResponse', getRecommendationsResponse_httpStatus - The response's http status code.

GetPersonalizedRanking

data GetPersonalizedRanking

See: newGetPersonalizedRanking smart constructor.

Constructors

GetPersonalizedRanking' (Maybe (HashMap Text (Sensitive Text))) (Maybe Text) (Maybe (HashMap Text (Sensitive Text))) Text [Text] Text

newGetPersonalizedRanking

Arguments

:: Text
$sel:campaignArn:GetPersonalizedRanking'

-> Text
$sel:userId:GetPersonalizedRanking'

-> GetPersonalizedRanking

Create a value of GetPersonalizedRanking with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:context:GetPersonalizedRanking', getPersonalizedRanking_context - The contextual metadata to use when getting recommendations. Contextual metadata includes any interaction information that might be relevant when getting a user's recommendations, such as the user's current location or device type.

$sel:filterArn:GetPersonalizedRanking', getPersonalizedRanking_filterArn - The Amazon Resource Name (ARN) of a filter you created to include items or exclude items from recommendations for a given user. For more information, see Filtering Recommendations.

$sel:filterValues:GetPersonalizedRanking', getPersonalizedRanking_filterValues - The values to use when filtering recommendations. For each placeholder parameter in your filter expression, provide the parameter name (in matching case) as a key and the filter value(s) as the corresponding value. Separate multiple values for one parameter with a comma.

For filter expressions that use an INCLUDE element to include items, you must provide values for all parameters that are defined in the expression. For filters with expressions that use an EXCLUDE element to exclude items, you can omit the filter-values.In this case, Amazon Personalize doesn't use that portion of the expression to filter recommendations.

For more information, see Filtering Recommendations.

$sel:campaignArn:GetPersonalizedRanking', getPersonalizedRanking_campaignArn - The Amazon Resource Name (ARN) of the campaign to use for generating the personalized ranking.

$sel:inputList:GetPersonalizedRanking', getPersonalizedRanking_inputList - A list of items (by itemId) to rank. If an item was not included in the training dataset, the item is appended to the end of the reranked list. The maximum is 500.

$sel:userId:GetPersonalizedRanking', getPersonalizedRanking_userId - The user for which you want the campaign to provide a personalized ranking.

data GetPersonalizedRankingResponse

See: newGetPersonalizedRankingResponse smart constructor.

Constructors

GetPersonalizedRankingResponse' (Maybe [PredictedItem]) (Maybe Text) Int

newGetPersonalizedRankingResponse

Arguments

:: Int
$sel:httpStatus:GetPersonalizedRankingResponse'

-> GetPersonalizedRankingResponse

Create a value of GetPersonalizedRankingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:personalizedRanking:GetPersonalizedRankingResponse', getPersonalizedRankingResponse_personalizedRanking - A list of items in order of most likely interest to the user. The maximum is 500.

$sel:recommendationId:GetPersonalizedRankingResponse', getPersonalizedRankingResponse_recommendationId - The ID of the recommendation.

$sel:httpStatus:GetPersonalizedRankingResponse', getPersonalizedRankingResponse_httpStatus - The response's http status code.

Types

PredictedItem

data PredictedItem

An object that identifies an item.

The and APIs return a list of PredictedItems.

See: newPredictedItem smart constructor.

Constructors

PredictedItem' (Maybe Double) (Maybe Text)

newPredictedItem :: PredictedItem

Create a value of PredictedItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:score:PredictedItem', predictedItem_score - A numeric representation of the model's certainty that the item will be the next user selection. For more information on scoring logic, see how-scores-work.

$sel:itemId:PredictedItem', predictedItem_itemId - The recommended item ID.