Network.AWS.PersonalizeRuntime.GetPersonalizedRanking

Description

Re-ranks a list of recommended items for the given user. The first item in the list is deemed the most likely item to be of interest to the user.

The solution backing the campaign must have been created using a recipe of type PERSONALIZED_RANKING.

Creating a Request

data GetPersonalizedRanking

See: newGetPersonalizedRanking smart constructor.

Constructors

GetPersonalizedRanking'

Fields

context :: Maybe (HashMap Text (Sensitive Text))

The contextual metadata to use when getting recommendations. Contextual metadata includes any interaction information that might be relevant when getting a user's recommendations, such as the user's current location or device type.

filterArn :: Maybe Text

The Amazon Resource Name (ARN) of a filter you created to include items or exclude items from recommendations for a given user. For more information, see Filtering Recommendations.

filterValues :: Maybe (HashMap Text (Sensitive Text))

The values to use when filtering recommendations. For each placeholder parameter in your filter expression, provide the parameter name (in matching case) as a key and the filter value(s) as the corresponding value. Separate multiple values for one parameter with a comma.

For filter expressions that use an INCLUDE element to include items, you must provide values for all parameters that are defined in the expression. For filters with expressions that use an EXCLUDE element to exclude items, you can omit the filter-values.In this case, Amazon Personalize doesn't use that portion of the expression to filter recommendations.

For more information, see Filtering Recommendations.

campaignArn :: Text

The Amazon Resource Name (ARN) of the campaign to use for generating the personalized ranking.

inputList :: [Text]

A list of items (by itemId) to rank. If an item was not included in the training dataset, the item is appended to the end of the reranked list. The maximum is 500.

userId :: Text

The user for which you want the campaign to provide a personalized ranking.

newGetPersonalizedRanking

Arguments

:: Text
$sel:campaignArn:GetPersonalizedRanking'

-> Text
$sel:userId:GetPersonalizedRanking'

-> GetPersonalizedRanking

Create a value of GetPersonalizedRanking with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:context:GetPersonalizedRanking', getPersonalizedRanking_context - The contextual metadata to use when getting recommendations. Contextual metadata includes any interaction information that might be relevant when getting a user's recommendations, such as the user's current location or device type.

$sel:filterArn:GetPersonalizedRanking', getPersonalizedRanking_filterArn - The Amazon Resource Name (ARN) of a filter you created to include items or exclude items from recommendations for a given user. For more information, see Filtering Recommendations.

$sel:filterValues:GetPersonalizedRanking', getPersonalizedRanking_filterValues - The values to use when filtering recommendations. For each placeholder parameter in your filter expression, provide the parameter name (in matching case) as a key and the filter value(s) as the corresponding value. Separate multiple values for one parameter with a comma.

For filter expressions that use an INCLUDE element to include items, you must provide values for all parameters that are defined in the expression. For filters with expressions that use an EXCLUDE element to exclude items, you can omit the filter-values.In this case, Amazon Personalize doesn't use that portion of the expression to filter recommendations.

For more information, see Filtering Recommendations.

$sel:campaignArn:GetPersonalizedRanking', getPersonalizedRanking_campaignArn - The Amazon Resource Name (ARN) of the campaign to use for generating the personalized ranking.

$sel:inputList:GetPersonalizedRanking', getPersonalizedRanking_inputList - A list of items (by itemId) to rank. If an item was not included in the training dataset, the item is appended to the end of the reranked list. The maximum is 500.

$sel:userId:GetPersonalizedRanking', getPersonalizedRanking_userId - The user for which you want the campaign to provide a personalized ranking.

Request Lenses

getPersonalizedRanking_context :: Lens' GetPersonalizedRanking (Maybe (HashMap Text Text))

The contextual metadata to use when getting recommendations. Contextual metadata includes any interaction information that might be relevant when getting a user's recommendations, such as the user's current location or device type.

getPersonalizedRanking_filterArn :: Lens' GetPersonalizedRanking (Maybe Text)

The Amazon Resource Name (ARN) of a filter you created to include items or exclude items from recommendations for a given user. For more information, see Filtering Recommendations.

getPersonalizedRanking_filterValues :: Lens' GetPersonalizedRanking (Maybe (HashMap Text Text))

The values to use when filtering recommendations. For each placeholder parameter in your filter expression, provide the parameter name (in matching case) as a key and the filter value(s) as the corresponding value. Separate multiple values for one parameter with a comma.

For filter expressions that use an INCLUDE element to include items, you must provide values for all parameters that are defined in the expression. For filters with expressions that use an EXCLUDE element to exclude items, you can omit the filter-values.In this case, Amazon Personalize doesn't use that portion of the expression to filter recommendations.

For more information, see Filtering Recommendations.

getPersonalizedRanking_campaignArn :: Lens' GetPersonalizedRanking Text

The Amazon Resource Name (ARN) of the campaign to use for generating the personalized ranking.

getPersonalizedRanking_inputList :: Lens' GetPersonalizedRanking [Text]

A list of items (by itemId) to rank. If an item was not included in the training dataset, the item is appended to the end of the reranked list. The maximum is 500.

getPersonalizedRanking_userId :: Lens' GetPersonalizedRanking Text

The user for which you want the campaign to provide a personalized ranking.

Destructuring the Response

data GetPersonalizedRankingResponse

See: newGetPersonalizedRankingResponse smart constructor.

Constructors

GetPersonalizedRankingResponse'

Fields

personalizedRanking :: Maybe [PredictedItem]

A list of items in order of most likely interest to the user. The maximum is 500.

recommendationId :: Maybe Text

The ID of the recommendation.

httpStatus :: Int

The response's http status code.

newGetPersonalizedRankingResponse

Arguments

:: Int
$sel:httpStatus:GetPersonalizedRankingResponse'

-> GetPersonalizedRankingResponse

Create a value of GetPersonalizedRankingResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:personalizedRanking:GetPersonalizedRankingResponse', getPersonalizedRankingResponse_personalizedRanking - A list of items in order of most likely interest to the user. The maximum is 500.

$sel:recommendationId:GetPersonalizedRankingResponse', getPersonalizedRankingResponse_recommendationId - The ID of the recommendation.

$sel:httpStatus:GetPersonalizedRankingResponse', getPersonalizedRankingResponse_httpStatus - The response's http status code.

Response Lenses

getPersonalizedRankingResponse_personalizedRanking :: Lens' GetPersonalizedRankingResponse (Maybe [PredictedItem])

A list of items in order of most likely interest to the user. The maximum is 500.

getPersonalizedRankingResponse_recommendationId :: Lens' GetPersonalizedRankingResponse (Maybe Text)

The ID of the recommendation.

getPersonalizedRankingResponse_httpStatus :: Lens' GetPersonalizedRankingResponse Int

The response's http status code.