Network.AWS.PersonalizeRuntime.Lens

Description

Operations

GetRecommendations

getRecommendations_context :: Lens' GetRecommendations (Maybe (HashMap Text Text))

The contextual metadata to use when getting recommendations. Contextual metadata includes any interaction information that might be relevant when getting a user's recommendations, such as the user's current location or device type.

getRecommendations_itemId :: Lens' GetRecommendations (Maybe Text)

The item ID to provide recommendations for.

Required for RELATED_ITEMS recipe type.

getRecommendations_userId :: Lens' GetRecommendations (Maybe Text)

The user ID to provide recommendations for.

Required for USER_PERSONALIZATION recipe type.

getRecommendations_numResults :: Lens' GetRecommendations (Maybe Natural)

The number of results to return. The default is 25. The maximum is 500.

getRecommendations_filterArn :: Lens' GetRecommendations (Maybe Text)

The ARN of the filter to apply to the returned recommendations. For more information, see Filtering Recommendations.

When using this parameter, be sure the filter resource is ACTIVE.

getRecommendations_filterValues :: Lens' GetRecommendations (Maybe (HashMap Text Text))

The values to use when filtering recommendations. For each placeholder parameter in your filter expression, provide the parameter name (in matching case) as a key and the filter value(s) as the corresponding value. Separate multiple values for one parameter with a comma.

For filter expressions that use an INCLUDE element to include items, you must provide values for all parameters that are defined in the expression. For filters with expressions that use an EXCLUDE element to exclude items, you can omit the filter-values.In this case, Amazon Personalize doesn't use that portion of the expression to filter recommendations.

For more information, see Filtering Recommendations.

getRecommendations_campaignArn :: Lens' GetRecommendations Text

The Amazon Resource Name (ARN) of the campaign to use for getting recommendations.

getRecommendationsResponse_recommendationId :: Lens' GetRecommendationsResponse (Maybe Text)

The ID of the recommendation.

getRecommendationsResponse_itemList :: Lens' GetRecommendationsResponse (Maybe [PredictedItem])

A list of recommendations sorted in ascending order by prediction score. There can be a maximum of 500 items in the list.

getRecommendationsResponse_httpStatus :: Lens' GetRecommendationsResponse Int

The response's http status code.

GetPersonalizedRanking

getPersonalizedRanking_context :: Lens' GetPersonalizedRanking (Maybe (HashMap Text Text))

The contextual metadata to use when getting recommendations. Contextual metadata includes any interaction information that might be relevant when getting a user's recommendations, such as the user's current location or device type.

getPersonalizedRanking_filterArn :: Lens' GetPersonalizedRanking (Maybe Text)

The Amazon Resource Name (ARN) of a filter you created to include items or exclude items from recommendations for a given user. For more information, see Filtering Recommendations.

getPersonalizedRanking_filterValues :: Lens' GetPersonalizedRanking (Maybe (HashMap Text Text))

The values to use when filtering recommendations. For each placeholder parameter in your filter expression, provide the parameter name (in matching case) as a key and the filter value(s) as the corresponding value. Separate multiple values for one parameter with a comma.

For filter expressions that use an INCLUDE element to include items, you must provide values for all parameters that are defined in the expression. For filters with expressions that use an EXCLUDE element to exclude items, you can omit the filter-values.In this case, Amazon Personalize doesn't use that portion of the expression to filter recommendations.

For more information, see Filtering Recommendations.

getPersonalizedRanking_campaignArn :: Lens' GetPersonalizedRanking Text

The Amazon Resource Name (ARN) of the campaign to use for generating the personalized ranking.

getPersonalizedRanking_inputList :: Lens' GetPersonalizedRanking [Text]

A list of items (by itemId) to rank. If an item was not included in the training dataset, the item is appended to the end of the reranked list. The maximum is 500.

getPersonalizedRanking_userId :: Lens' GetPersonalizedRanking Text

The user for which you want the campaign to provide a personalized ranking.

getPersonalizedRankingResponse_personalizedRanking :: Lens' GetPersonalizedRankingResponse (Maybe [PredictedItem])

A list of items in order of most likely interest to the user. The maximum is 500.

getPersonalizedRankingResponse_recommendationId :: Lens' GetPersonalizedRankingResponse (Maybe Text)

The ID of the recommendation.

getPersonalizedRankingResponse_httpStatus :: Lens' GetPersonalizedRankingResponse Int

The response's http status code.

Types

PredictedItem

predictedItem_score :: Lens' PredictedItem (Maybe Double)

A numeric representation of the model's certainty that the item will be the next user selection. For more information on scoring logic, see how-scores-work.

predictedItem_itemId :: Lens' PredictedItem (Maybe Text)

The recommended item ID.