Network.AWS.Pi

Description

Derived from API version 2018-02-27 of the AWS service descriptions, licensed under Apache 2.0.

Amazon RDS Performance Insights

Amazon RDS Performance Insights enables you to monitor and explore different dimensions of database load based on data captured from a running DB instance. The guide provides detailed information about Performance Insights data types, parameters and errors.

When Performance Insights is enabled, the Amazon RDS Performance Insights API provides visibility into the performance of your DB instance. Amazon CloudWatch provides the authoritative source for AWS service-vended monitoring metrics. Performance Insights offers a domain-specific view of DB load.

DB load is measured as Average Active Sessions. Performance Insights provides the data to API consumers as a two-dimensional time-series dataset. The time dimension provides DB load data for each time point in the queried time range. Each time point decomposes overall load in relation to the requested dimensions, measured at that time point. Examples include SQL, Wait event, User, and Host.

To learn more about Performance Insights and Amazon Aurora DB instances, go to the Amazon Aurora User Guide.

To learn more about Performance Insights and Amazon RDS DB instances, go to the Amazon RDS User Guide.

Service Configuration

defaultService :: Service

API version 2018-02-27 of the Amazon Performance Insights SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by Pi.

InvalidArgumentException

_InvalidArgumentException :: AsError a => Getting (First ServiceError) a ServiceError

One of the arguments provided is invalid for this request.

NotAuthorizedException

_NotAuthorizedException :: AsError a => Getting (First ServiceError) a ServiceError

The user is not authorized to perform this request.

InternalServiceError

_InternalServiceError :: AsError a => Getting (First ServiceError) a ServiceError

The request failed due to an unknown error.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

GetDimensionKeyDetails

data GetDimensionKeyDetails

See: newGetDimensionKeyDetails smart constructor.

Constructors

GetDimensionKeyDetails' (Maybe (NonEmpty Text)) ServiceType Text Text Text

newGetDimensionKeyDetails

Arguments

:: ServiceType
$sel:serviceType:GetDimensionKeyDetails'

-> Text
$sel:identifier:GetDimensionKeyDetails'

-> Text
$sel:group':GetDimensionKeyDetails'

-> Text
$sel:groupIdentifier:GetDimensionKeyDetails'

-> GetDimensionKeyDetails

Create a value of GetDimensionKeyDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestedDimensions:GetDimensionKeyDetails', getDimensionKeyDetails_requestedDimensions - A list of dimensions to retrieve the detail data for within the given dimension group. For the dimension group db.sql, specify either the full dimension name db.sql.statement or the short dimension name statement. If you don't specify this parameter, Performance Insights returns all dimension data within the specified dimension group.

$sel:serviceType:GetDimensionKeyDetails', getDimensionKeyDetails_serviceType - The AWS service for which Performance Insights returns data. The only valid value is RDS.

$sel:identifier:GetDimensionKeyDetails', getDimensionKeyDetails_identifier - The ID for a data source from which to gather dimension data. This ID must be immutable and unique within an AWS Region. When a DB instance is the data source, specify its DbiResourceId value. For example, specify db-ABCDEFGHIJKLMNOPQRSTU1VW2X.

$sel:group':GetDimensionKeyDetails', getDimensionKeyDetails_group - The name of the dimension group. The only valid value is db.sql. Performance Insights searches the specified group for the dimension group ID.

$sel:groupIdentifier:GetDimensionKeyDetails', getDimensionKeyDetails_groupIdentifier - The ID of the dimension group from which to retrieve dimension details. For dimension group db.sql, the group ID is db.sql.id.

data GetDimensionKeyDetailsResponse

See: newGetDimensionKeyDetailsResponse smart constructor.

Constructors

GetDimensionKeyDetailsResponse' (Maybe [DimensionKeyDetail]) Int

newGetDimensionKeyDetailsResponse

Arguments

:: Int
$sel:httpStatus:GetDimensionKeyDetailsResponse'

-> GetDimensionKeyDetailsResponse

Create a value of GetDimensionKeyDetailsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dimensions:GetDimensionKeyDetailsResponse', getDimensionKeyDetailsResponse_dimensions - The details for the requested dimensions.

$sel:httpStatus:GetDimensionKeyDetailsResponse', getDimensionKeyDetailsResponse_httpStatus - The response's http status code.

GetResourceMetrics

data GetResourceMetrics

See: newGetResourceMetrics smart constructor.

Constructors

GetResourceMetrics' (Maybe Int) (Maybe Text) (Maybe Natural) ServiceType Text (NonEmpty MetricQuery) POSIX POSIX

newGetResourceMetrics

Arguments

:: ServiceType
$sel:serviceType:GetResourceMetrics'

-> Text
$sel:identifier:GetResourceMetrics'

-> NonEmpty MetricQuery
$sel:metricQueries:GetResourceMetrics'

-> UTCTime
$sel:startTime:GetResourceMetrics'

-> UTCTime
$sel:endTime:GetResourceMetrics'

-> GetResourceMetrics

Create a value of GetResourceMetrics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:periodInSeconds:GetResourceMetrics', getResourceMetrics_periodInSeconds - The granularity, in seconds, of the data points returned from Performance Insights. A period can be as short as one second, or as long as one day (86400 seconds). Valid values are:

1 (one second)

60 (one minute)

300 (five minutes)

3600 (one hour)

86400 (twenty-four hours)

If you don't specify PeriodInSeconds, then Performance Insights will choose a value for you, with a goal of returning roughly 100-200 data points in the response.

$sel:nextToken:GetResourceMetrics', getResourceMetrics_nextToken - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the token, up to the value specified by MaxRecords.

$sel:maxResults:GetResourceMetrics', getResourceMetrics_maxResults - The maximum number of items to return in the response. If more items exist than the specified MaxRecords value, a pagination token is included in the response so that the remaining results can be retrieved.

$sel:serviceType:GetResourceMetrics', getResourceMetrics_serviceType - The AWS service for which Performance Insights returns metrics. The only valid value for ServiceType is RDS.

$sel:identifier:GetResourceMetrics', getResourceMetrics_identifier - An immutable, AWS Region-unique identifier for a data source. Performance Insights gathers metrics from this data source.

To use a DB instance as a data source, specify its DbiResourceId value. For example, specify db-FAIHNTYBKTGAUSUZQYPDS2GW4A.

$sel:metricQueries:GetResourceMetrics', getResourceMetrics_metricQueries - An array of one or more queries to perform. Each query must specify a Performance Insights metric, and can optionally specify aggregation and filtering criteria.

$sel:startTime:GetResourceMetrics', getResourceMetrics_startTime - The date and time specifying the beginning of the requested time series data. You can't specify a StartTime that's earlier than 7 days ago. The value specified is inclusive - data points equal to or greater than StartTime will be returned.

The value for StartTime must be earlier than the value for EndTime.

$sel:endTime:GetResourceMetrics', getResourceMetrics_endTime - The date and time specifying the end of the requested time series data. The value specified is exclusive - data points less than (but not equal to) EndTime will be returned.

The value for EndTime must be later than the value for StartTime.

data GetResourceMetricsResponse

See: newGetResourceMetricsResponse smart constructor.

Constructors

GetResourceMetricsResponse' (Maybe Text) (Maybe [MetricKeyDataPoints]) (Maybe POSIX) (Maybe POSIX) (Maybe Text) Int

newGetResourceMetricsResponse

Arguments

:: Int
$sel:httpStatus:GetResourceMetricsResponse'

-> GetResourceMetricsResponse

Create a value of GetResourceMetricsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:identifier:GetResourceMetrics', getResourceMetricsResponse_identifier - An immutable, AWS Region-unique identifier for a data source. Performance Insights gathers metrics from this data source.

To use a DB instance as a data source, you specify its DbiResourceId value - for example: db-FAIHNTYBKTGAUSUZQYPDS2GW4A

$sel:metricList:GetResourceMetricsResponse', getResourceMetricsResponse_metricList - An array of metric results,, where each array element contains all of the data points for a particular dimension.

$sel:alignedEndTime:GetResourceMetricsResponse', getResourceMetricsResponse_alignedEndTime - The end time for the returned metrics, after alignment to a granular boundary (as specified by PeriodInSeconds). AlignedEndTime will be greater than or equal to the value of the user-specified Endtime.

$sel:alignedStartTime:GetResourceMetricsResponse', getResourceMetricsResponse_alignedStartTime - The start time for the returned metrics, after alignment to a granular boundary (as specified by PeriodInSeconds). AlignedStartTime will be less than or equal to the value of the user-specified StartTime.

$sel:nextToken:GetResourceMetrics', getResourceMetricsResponse_nextToken - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the token, up to the value specified by MaxRecords.

$sel:httpStatus:GetResourceMetricsResponse', getResourceMetricsResponse_httpStatus - The response's http status code.

DescribeDimensionKeys

data DescribeDimensionKeys

See: newDescribeDimensionKeys smart constructor.

Constructors

DescribeDimensionKeys' (Maybe Int) (Maybe Text) (Maybe (HashMap Text Text)) (Maybe Natural) (Maybe DimensionGroup) ServiceType Text POSIX POSIX Text DimensionGroup

newDescribeDimensionKeys

Arguments

:: ServiceType
$sel:serviceType:DescribeDimensionKeys'

-> Text
$sel:identifier:DescribeDimensionKeys'

-> UTCTime
$sel:startTime:DescribeDimensionKeys'

-> UTCTime
$sel:endTime:DescribeDimensionKeys'

-> Text
$sel:metric:DescribeDimensionKeys'

-> DimensionGroup
$sel:groupBy:DescribeDimensionKeys'

-> DescribeDimensionKeys

Create a value of DescribeDimensionKeys with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:periodInSeconds:DescribeDimensionKeys', describeDimensionKeys_periodInSeconds - The granularity, in seconds, of the data points returned from Performance Insights. A period can be as short as one second, or as long as one day (86400 seconds). Valid values are:

1 (one second)

60 (one minute)

300 (five minutes)

3600 (one hour)

86400 (twenty-four hours)

If you don't specify PeriodInSeconds, then Performance Insights chooses a value for you, with a goal of returning roughly 100-200 data points in the response.

$sel:nextToken:DescribeDimensionKeys', describeDimensionKeys_nextToken - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the token, up to the value specified by MaxRecords.

$sel:filter':DescribeDimensionKeys', describeDimensionKeys_filter - One or more filters to apply in the request. Restrictions:

Any number of filters by the same dimension, as specified in the GroupBy or Partition parameters.

A single filter for any other dimension in this dimension group.

$sel:maxResults:DescribeDimensionKeys', describeDimensionKeys_maxResults - The maximum number of items to return in the response. If more items exist than the specified MaxRecords value, a pagination token is included in the response so that the remaining results can be retrieved.

$sel:partitionBy:DescribeDimensionKeys', describeDimensionKeys_partitionBy - For each dimension specified in GroupBy, specify a secondary dimension to further subdivide the partition keys in the response.

$sel:serviceType:DescribeDimensionKeys', describeDimensionKeys_serviceType - The AWS service for which Performance Insights will return metrics. The only valid value for ServiceType is RDS.

$sel:identifier:DescribeDimensionKeys', describeDimensionKeys_identifier - An immutable, AWS Region-unique identifier for a data source. Performance Insights gathers metrics from this data source.

To use an Amazon RDS instance as a data source, you specify its DbiResourceId value. For example, specify db-FAIHNTYBKTGAUSUZQYPDS2GW4A

$sel:startTime:DescribeDimensionKeys', describeDimensionKeys_startTime - The date and time specifying the beginning of the requested time series data. You must specify a StartTime within the past 7 days. The value specified is inclusive, which means that data points equal to or greater than StartTime are returned.

The value for StartTime must be earlier than the value for EndTime.

$sel:endTime:DescribeDimensionKeys', describeDimensionKeys_endTime - The date and time specifying the end of the requested time series data. The value specified is exclusive, which means that data points less than (but not equal to) EndTime are returned.

The value for EndTime must be later than the value for StartTime.

$sel:metric:DescribeDimensionKeys', describeDimensionKeys_metric - The name of a Performance Insights metric to be measured.

Valid values for Metric are:

db.load.avg - a scaled representation of the number of active sessions for the database engine.

db.sampledload.avg - the raw number of active sessions for the database engine.

If the number of active sessions is less than an internal Performance Insights threshold, db.load.avg and db.sampledload.avg are the same value. If the number of active sessions is greater than the internal threshold, Performance Insights samples the active sessions, with db.load.avg showing the scaled values, db.sampledload.avg showing the raw values, and db.sampledload.avg less than db.load.avg. For most use cases, you can query db.load.avg only.

$sel:groupBy:DescribeDimensionKeys', describeDimensionKeys_groupBy - A specification for how to aggregate the data points from a query result. You must specify a valid dimension group. Performance Insights returns all dimensions within this group, unless you provide the names of specific dimensions within this group. You can also request that Performance Insights return a limited number of values for a dimension.

data DescribeDimensionKeysResponse

See: newDescribeDimensionKeysResponse smart constructor.

Constructors

DescribeDimensionKeysResponse' (Maybe POSIX) (Maybe POSIX) (Maybe [DimensionKeyDescription]) (Maybe Text) (Maybe [ResponsePartitionKey]) Int

newDescribeDimensionKeysResponse

Arguments

:: Int
$sel:httpStatus:DescribeDimensionKeysResponse'

-> DescribeDimensionKeysResponse

Create a value of DescribeDimensionKeysResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:alignedEndTime:DescribeDimensionKeysResponse', describeDimensionKeysResponse_alignedEndTime - The end time for the returned dimension keys, after alignment to a granular boundary (as specified by PeriodInSeconds). AlignedEndTime will be greater than or equal to the value of the user-specified Endtime.

$sel:alignedStartTime:DescribeDimensionKeysResponse', describeDimensionKeysResponse_alignedStartTime - The start time for the returned dimension keys, after alignment to a granular boundary (as specified by PeriodInSeconds). AlignedStartTime will be less than or equal to the value of the user-specified StartTime.

$sel:keys:DescribeDimensionKeysResponse', describeDimensionKeysResponse_keys - The dimension keys that were requested.

$sel:nextToken:DescribeDimensionKeys', describeDimensionKeysResponse_nextToken - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the token, up to the value specified by MaxRecords.

$sel:partitionKeys:DescribeDimensionKeysResponse', describeDimensionKeysResponse_partitionKeys - If PartitionBy was present in the request, PartitionKeys contains the breakdown of dimension keys by the specified partitions.

$sel:httpStatus:DescribeDimensionKeysResponse', describeDimensionKeysResponse_httpStatus - The response's http status code.

Types

DetailStatus

newtype DetailStatus

Constructors

DetailStatus'

Fields

fromDetailStatus :: Text

Bundled Patterns

pattern DetailStatus_AVAILABLE :: DetailStatus

pattern DetailStatus_PROCESSING :: DetailStatus

pattern DetailStatus_UNAVAILABLE :: DetailStatus

ServiceType

newtype ServiceType

Constructors

ServiceType'

Fields

fromServiceType :: Text

Bundled Patterns

pattern ServiceType_RDS :: ServiceType

DataPoint

data DataPoint

A timestamp, and a single numerical value, which together represent a measurement at a particular point in time.

See: newDataPoint smart constructor.

Constructors

DataPoint' POSIX Double

newDataPoint

Arguments

:: UTCTime
$sel:timestamp:DataPoint'

-> Double
$sel:value:DataPoint'

-> DataPoint

Create a value of DataPoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timestamp:DataPoint', dataPoint_timestamp - The time, in epoch format, associated with a particular Value.

$sel:value:DataPoint', dataPoint_value - The actual value associated with a particular Timestamp.

DimensionGroup

data DimensionGroup

A logical grouping of Performance Insights metrics for a related subject area. For example, the db.sql dimension group consists of the following dimensions: db.sql.id, db.sql.db_id, db.sql.statement, and db.sql.tokenized_id.

Each response element returns a maximum of 500 bytes. For larger elements, such as SQL statements, only the first 500 bytes are returned.

See: newDimensionGroup smart constructor.

Constructors

DimensionGroup' (Maybe Natural) (Maybe (NonEmpty Text)) Text

newDimensionGroup

Arguments

:: Text
$sel:group':DimensionGroup'

-> DimensionGroup

Create a value of DimensionGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:limit:DimensionGroup', dimensionGroup_limit - The maximum number of items to fetch for this dimension group.

$sel:dimensions:DimensionGroup', dimensionGroup_dimensions - A list of specific dimensions from a dimension group. If this parameter is not present, then it signifies that all of the dimensions in the group were requested, or are present in the response.

Valid values for elements in the Dimensions array are:

db.application.name - The name of the application that is connected to the database (only Aurora PostgreSQL and RDS PostgreSQL)

db.host.id - The host ID of the connected client (all engines)

db.host.name - The host name of the connected client (all engines)

db.name - The name of the database to which the client is connected (only Aurora PostgreSQL, RDS PostgreSQL, Aurora MySQL, RDS MySQL, and MariaDB)

db.session_type.name - The type of the current session (only Aurora PostgreSQL and RDS PostgreSQL)

db.sql.id - The SQL ID generated by Performance Insights (all engines)

db.sql.db_id - The SQL ID generated by the database (all engines)

db.sql.statement - The SQL text that is being executed (all engines)

db.sql.tokenized_id

db.sql_tokenized.id - The SQL digest ID generated by Performance Insights (all engines)

db.sql_tokenized.db_id - SQL digest ID generated by the database (all engines)

db.sql_tokenized.statement - The SQL digest text (all engines)

db.user.id - The ID of the user logged in to the database (all engines)

db.user.name - The name of the user logged in to the database (all engines)

db.wait_event.name - The event for which the backend is waiting (all engines)

db.wait_event.type - The type of event for which the backend is waiting (all engines)

db.wait_event_type.name - The name of the event type for which the backend is waiting (all engines)

$sel:group':DimensionGroup', dimensionGroup_group - The name of the dimension group. Valid values are:

db - The name of the database to which the client is connected (only Aurora PostgreSQL, RDS PostgreSQL, Aurora MySQL, RDS MySQL, and MariaDB)

db.application - The name of the application that is connected to the database (only Aurora PostgreSQL and RDS PostgreSQL)

db.host - The host name of the connected client (all engines)

db.session_type - The type of the current session (only Aurora PostgreSQL and RDS PostgreSQL)

db.sql - The SQL that is currently executing (all engines)

db.sql_tokenized - The SQL digest (all engines)

db.wait_event - The event for which the database backend is waiting (all engines)

db.wait_event_type - The type of event for which the database backend is waiting (all engines)

db.user - The user logged in to the database (all engines)

DimensionKeyDescription

data DimensionKeyDescription

An array of descriptions and aggregated values for each dimension within a dimension group.

See: newDimensionKeyDescription smart constructor.

Constructors

DimensionKeyDescription' (Maybe [Double]) (Maybe Double) (Maybe (HashMap Text Text))

newDimensionKeyDescription :: DimensionKeyDescription

Create a value of DimensionKeyDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:partitions:DimensionKeyDescription', dimensionKeyDescription_partitions - If PartitionBy was specified, PartitionKeys contains the dimensions that were.

$sel:total:DimensionKeyDescription', dimensionKeyDescription_total - The aggregated metric value for the dimension(s), over the requested time range.

$sel:dimensions:DimensionKeyDescription', dimensionKeyDescription_dimensions - A map of name-value pairs for the dimensions in the group.

DimensionKeyDetail

data DimensionKeyDetail

An object that describes the details for a specified dimension.

See: newDimensionKeyDetail smart constructor.

Constructors

DimensionKeyDetail' (Maybe DetailStatus) (Maybe Text) (Maybe Text)

newDimensionKeyDetail :: DimensionKeyDetail

Create a value of DimensionKeyDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DimensionKeyDetail', dimensionKeyDetail_status - The status of the dimension detail data. Possible values include the following:

AVAILABLE - The dimension detail data is ready to be retrieved.

PROCESSING - The dimension detail data isn't ready to be retrieved because more processing time is required. If the requested detail data for db.sql.statement has the status PROCESSING, Performance Insights returns the truncated query.

UNAVAILABLE - The dimension detail data could not be collected successfully.

$sel:dimension:DimensionKeyDetail', dimensionKeyDetail_dimension - The full name of the dimension. The full name includes the group name and key name. The only valid value is db.sql.statement.

$sel:value:DimensionKeyDetail', dimensionKeyDetail_value - The value of the dimension detail data. For the db.sql.statement dimension, this value is either the full or truncated SQL query, depending on the return status.

MetricKeyDataPoints

data MetricKeyDataPoints

A time-ordered series of data points, corresponding to a dimension of a Performance Insights metric.

See: newMetricKeyDataPoints smart constructor.

Constructors

MetricKeyDataPoints' (Maybe [DataPoint]) (Maybe ResponseResourceMetricKey)

newMetricKeyDataPoints :: MetricKeyDataPoints

Create a value of MetricKeyDataPoints with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataPoints:MetricKeyDataPoints', metricKeyDataPoints_dataPoints - An array of timestamp-value pairs, representing measurements over a period of time.

$sel:key:MetricKeyDataPoints', metricKeyDataPoints_key - The dimension(s) to which the data points apply.

MetricQuery

data MetricQuery

A single query to be processed. You must provide the metric to query. If no other parameters are specified, Performance Insights returns all of the data points for that metric. You can optionally request that the data points be aggregated by dimension group ( GroupBy), and return only those data points that match your criteria (Filter).

See: newMetricQuery smart constructor.

Constructors

MetricQuery' (Maybe DimensionGroup) (Maybe (HashMap Text Text)) Text

newMetricQuery

Arguments

:: Text
$sel:metric:MetricQuery'

-> MetricQuery

Create a value of MetricQuery with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:groupBy:MetricQuery', metricQuery_groupBy - A specification for how to aggregate the data points from a query result. You must specify a valid dimension group. Performance Insights will return all of the dimensions within that group, unless you provide the names of specific dimensions within that group. You can also request that Performance Insights return a limited number of values for a dimension.

$sel:filter':MetricQuery', metricQuery_filter - One or more filters to apply in the request. Restrictions:

Any number of filters by the same dimension, as specified in the GroupBy parameter.

A single filter for any other dimension in this dimension group.

$sel:metric:MetricQuery', metricQuery_metric - The name of a Performance Insights metric to be measured.

Valid values for Metric are:

db.load.avg - a scaled representation of the number of active sessions for the database engine.

db.sampledload.avg - the raw number of active sessions for the database engine.

If the number of active sessions is less than an internal Performance Insights threshold, db.load.avg and db.sampledload.avg are the same value. If the number of active sessions is greater than the internal threshold, Performance Insights samples the active sessions, with db.load.avg showing the scaled values, db.sampledload.avg showing the raw values, and db.sampledload.avg less than db.load.avg. For most use cases, you can query db.load.avg only.

ResponsePartitionKey

data ResponsePartitionKey

If PartitionBy was specified in a DescribeDimensionKeys request, the dimensions are returned in an array. Each element in the array specifies one dimension.

See: newResponsePartitionKey smart constructor.

Constructors

ResponsePartitionKey' (HashMap Text Text)

newResponsePartitionKey :: ResponsePartitionKey

Create a value of ResponsePartitionKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dimensions:ResponsePartitionKey', responsePartitionKey_dimensions - A dimension map that contains the dimension(s) for this partition.

ResponseResourceMetricKey

data ResponseResourceMetricKey

An object describing a Performance Insights metric and one or more dimensions for that metric.

See: newResponseResourceMetricKey smart constructor.

Constructors

ResponseResourceMetricKey' (Maybe (HashMap Text Text)) Text

newResponseResourceMetricKey

Arguments

:: Text
$sel:metric:ResponseResourceMetricKey'

-> ResponseResourceMetricKey

Create a value of ResponseResourceMetricKey with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dimensions:ResponseResourceMetricKey', responseResourceMetricKey_dimensions - The valid dimensions for the metric.

$sel:metric:ResponseResourceMetricKey', responseResourceMetricKey_metric - The name of a Performance Insights metric to be measured.

Valid values for Metric are:

db.load.avg - a scaled representation of the number of active sessions for the database engine.

db.sampledload.avg - the raw number of active sessions for the database engine.

If the number of active sessions is less than an internal Performance Insights threshold, db.load.avg and db.sampledload.avg are the same value. If the number of active sessions is greater than the internal threshold, Performance Insights samples the active sessions, with db.load.avg showing the scaled values, db.sampledload.avg showing the raw values, and db.sampledload.avg less than db.load.avg. For most use cases, you can query db.load.avg only.