Network.AWS.Pi.GetDimensionKeyDetails

Description

Get the attributes of the specified dimension group for a DB instance or data source. For example, if you specify a SQL ID, GetDimensionKeyDetails retrieves the full text of the dimension db.sql.statement associated with this ID. This operation is useful because GetResourceMetrics and DescribeDimensionKeys don't support retrieval of large SQL statement text.

Creating a Request

data GetDimensionKeyDetails

See: newGetDimensionKeyDetails smart constructor.

Constructors

GetDimensionKeyDetails'

Fields

requestedDimensions :: Maybe (NonEmpty Text)

A list of dimensions to retrieve the detail data for within the given dimension group. For the dimension group db.sql, specify either the full dimension name db.sql.statement or the short dimension name statement. If you don't specify this parameter, Performance Insights returns all dimension data within the specified dimension group.

serviceType :: ServiceType

The AWS service for which Performance Insights returns data. The only valid value is RDS.

identifier :: Text

The ID for a data source from which to gather dimension data. This ID must be immutable and unique within an AWS Region. When a DB instance is the data source, specify its DbiResourceId value. For example, specify db-ABCDEFGHIJKLMNOPQRSTU1VW2X.

group' :: Text

The name of the dimension group. The only valid value is db.sql. Performance Insights searches the specified group for the dimension group ID.

groupIdentifier :: Text

The ID of the dimension group from which to retrieve dimension details. For dimension group db.sql, the group ID is db.sql.id.

newGetDimensionKeyDetails

Arguments

:: ServiceType
$sel:serviceType:GetDimensionKeyDetails'

-> Text
$sel:identifier:GetDimensionKeyDetails'

-> Text
$sel:group':GetDimensionKeyDetails'

-> Text
$sel:groupIdentifier:GetDimensionKeyDetails'

-> GetDimensionKeyDetails

Create a value of GetDimensionKeyDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestedDimensions:GetDimensionKeyDetails', getDimensionKeyDetails_requestedDimensions - A list of dimensions to retrieve the detail data for within the given dimension group. For the dimension group db.sql, specify either the full dimension name db.sql.statement or the short dimension name statement. If you don't specify this parameter, Performance Insights returns all dimension data within the specified dimension group.

$sel:serviceType:GetDimensionKeyDetails', getDimensionKeyDetails_serviceType - The AWS service for which Performance Insights returns data. The only valid value is RDS.

$sel:identifier:GetDimensionKeyDetails', getDimensionKeyDetails_identifier - The ID for a data source from which to gather dimension data. This ID must be immutable and unique within an AWS Region. When a DB instance is the data source, specify its DbiResourceId value. For example, specify db-ABCDEFGHIJKLMNOPQRSTU1VW2X.

$sel:group':GetDimensionKeyDetails', getDimensionKeyDetails_group - The name of the dimension group. The only valid value is db.sql. Performance Insights searches the specified group for the dimension group ID.

$sel:groupIdentifier:GetDimensionKeyDetails', getDimensionKeyDetails_groupIdentifier - The ID of the dimension group from which to retrieve dimension details. For dimension group db.sql, the group ID is db.sql.id.

Request Lenses

getDimensionKeyDetails_requestedDimensions :: Lens' GetDimensionKeyDetails (Maybe (NonEmpty Text))

A list of dimensions to retrieve the detail data for within the given dimension group. For the dimension group db.sql, specify either the full dimension name db.sql.statement or the short dimension name statement. If you don't specify this parameter, Performance Insights returns all dimension data within the specified dimension group.

getDimensionKeyDetails_serviceType :: Lens' GetDimensionKeyDetails ServiceType

The AWS service for which Performance Insights returns data. The only valid value is RDS.

getDimensionKeyDetails_identifier :: Lens' GetDimensionKeyDetails Text

The ID for a data source from which to gather dimension data. This ID must be immutable and unique within an AWS Region. When a DB instance is the data source, specify its DbiResourceId value. For example, specify db-ABCDEFGHIJKLMNOPQRSTU1VW2X.

getDimensionKeyDetails_group :: Lens' GetDimensionKeyDetails Text

The name of the dimension group. The only valid value is db.sql. Performance Insights searches the specified group for the dimension group ID.

getDimensionKeyDetails_groupIdentifier :: Lens' GetDimensionKeyDetails Text

The ID of the dimension group from which to retrieve dimension details. For dimension group db.sql, the group ID is db.sql.id.

Destructuring the Response

data GetDimensionKeyDetailsResponse

See: newGetDimensionKeyDetailsResponse smart constructor.

Constructors

GetDimensionKeyDetailsResponse'

Fields

dimensions :: Maybe [DimensionKeyDetail]

The details for the requested dimensions.

httpStatus :: Int

The response's http status code.

newGetDimensionKeyDetailsResponse

Arguments

:: Int
$sel:httpStatus:GetDimensionKeyDetailsResponse'

-> GetDimensionKeyDetailsResponse

Create a value of GetDimensionKeyDetailsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dimensions:GetDimensionKeyDetailsResponse', getDimensionKeyDetailsResponse_dimensions - The details for the requested dimensions.

$sel:httpStatus:GetDimensionKeyDetailsResponse', getDimensionKeyDetailsResponse_httpStatus - The response's http status code.

Response Lenses

getDimensionKeyDetailsResponse_dimensions :: Lens' GetDimensionKeyDetailsResponse (Maybe [DimensionKeyDetail])

The details for the requested dimensions.

getDimensionKeyDetailsResponse_httpStatus :: Lens' GetDimensionKeyDetailsResponse Int

The response's http status code.