Network.AWS.Pi.Types.DimensionKeyDetail

Description

Documentation

data DimensionKeyDetail

An object that describes the details for a specified dimension.

See: newDimensionKeyDetail smart constructor.

Constructors

DimensionKeyDetail'

Fields

status :: Maybe DetailStatus

The status of the dimension detail data. Possible values include the following:

AVAILABLE - The dimension detail data is ready to be retrieved.

PROCESSING - The dimension detail data isn't ready to be retrieved because more processing time is required. If the requested detail data for db.sql.statement has the status PROCESSING, Performance Insights returns the truncated query.

UNAVAILABLE - The dimension detail data could not be collected successfully.

dimension :: Maybe Text

The full name of the dimension. The full name includes the group name and key name. The only valid value is db.sql.statement.

value :: Maybe Text

The value of the dimension detail data. For the db.sql.statement dimension, this value is either the full or truncated SQL query, depending on the return status.

newDimensionKeyDetail :: DimensionKeyDetail

Create a value of DimensionKeyDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DimensionKeyDetail', dimensionKeyDetail_status - The status of the dimension detail data. Possible values include the following:

AVAILABLE - The dimension detail data is ready to be retrieved.

PROCESSING - The dimension detail data isn't ready to be retrieved because more processing time is required. If the requested detail data for db.sql.statement has the status PROCESSING, Performance Insights returns the truncated query.

UNAVAILABLE - The dimension detail data could not be collected successfully.

$sel:dimension:DimensionKeyDetail', dimensionKeyDetail_dimension - The full name of the dimension. The full name includes the group name and key name. The only valid value is db.sql.statement.

$sel:value:DimensionKeyDetail', dimensionKeyDetail_value - The value of the dimension detail data. For the db.sql.statement dimension, this value is either the full or truncated SQL query, depending on the return status.

dimensionKeyDetail_status :: Lens' DimensionKeyDetail (Maybe DetailStatus)

The status of the dimension detail data. Possible values include the following:

AVAILABLE - The dimension detail data is ready to be retrieved.

PROCESSING - The dimension detail data isn't ready to be retrieved because more processing time is required. If the requested detail data for db.sql.statement has the status PROCESSING, Performance Insights returns the truncated query.

UNAVAILABLE - The dimension detail data could not be collected successfully.

dimensionKeyDetail_dimension :: Lens' DimensionKeyDetail (Maybe Text)

The full name of the dimension. The full name includes the group name and key name. The only valid value is db.sql.statement.

dimensionKeyDetail_value :: Lens' DimensionKeyDetail (Maybe Text)

The value of the dimension detail data. For the db.sql.statement dimension, this value is either the full or truncated SQL query, depending on the return status.