Network.AWS.Pinpoint.CreateInAppTemplate

Description

Creates a new message template for messages using the in-app message channel.

Creating a Request

data CreateInAppTemplate

See: newCreateInAppTemplate smart constructor.

Constructors

CreateInAppTemplate'

Fields

templateName :: Text

The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

inAppTemplateRequest :: InAppTemplateRequest

newCreateInAppTemplate

Arguments

:: Text
$sel:templateName:CreateInAppTemplate'

-> InAppTemplateRequest
$sel:inAppTemplateRequest:CreateInAppTemplate'

-> CreateInAppTemplate

Create a value of CreateInAppTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:CreateInAppTemplate', createInAppTemplate_templateName - The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

$sel:inAppTemplateRequest:CreateInAppTemplate', createInAppTemplate_inAppTemplateRequest - Undocumented member.

Request Lenses

createInAppTemplate_templateName :: Lens' CreateInAppTemplate Text

The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

createInAppTemplate_inAppTemplateRequest :: Lens' CreateInAppTemplate InAppTemplateRequest

Undocumented member.

Destructuring the Response

data CreateInAppTemplateResponse

See: newCreateInAppTemplateResponse smart constructor.

Constructors

CreateInAppTemplateResponse'

Fields

httpStatus :: Int

The response's http status code.

templateCreateMessageBody :: TemplateCreateMessageBody

newCreateInAppTemplateResponse

Arguments

:: Int
$sel:httpStatus:CreateInAppTemplateResponse'

-> TemplateCreateMessageBody
$sel:templateCreateMessageBody:CreateInAppTemplateResponse'

-> CreateInAppTemplateResponse

Create a value of CreateInAppTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateInAppTemplateResponse', createInAppTemplateResponse_httpStatus - The response's http status code.

$sel:templateCreateMessageBody:CreateInAppTemplateResponse', createInAppTemplateResponse_templateCreateMessageBody - Undocumented member.

Response Lenses

createInAppTemplateResponse_httpStatus :: Lens' CreateInAppTemplateResponse Int

The response's http status code.

createInAppTemplateResponse_templateCreateMessageBody :: Lens' CreateInAppTemplateResponse TemplateCreateMessageBody

Undocumented member.