Network.AWS.Pinpoint.CreateJourney

Description

Creates a journey for an application.

Creating a Request

data CreateJourney

See: newCreateJourney smart constructor.

Constructors

CreateJourney'

Fields

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

writeJourneyRequest :: WriteJourneyRequest

newCreateJourney

Arguments

:: Text
$sel:applicationId:CreateJourney'

-> WriteJourneyRequest
$sel:writeJourneyRequest:CreateJourney'

-> CreateJourney

Create a value of CreateJourney with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:CreateJourney', createJourney_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

$sel:writeJourneyRequest:CreateJourney', createJourney_writeJourneyRequest - Undocumented member.

Request Lenses

createJourney_applicationId :: Lens' CreateJourney Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

createJourney_writeJourneyRequest :: Lens' CreateJourney WriteJourneyRequest

Undocumented member.

Destructuring the Response

data CreateJourneyResponse

See: newCreateJourneyResponse smart constructor.

Constructors

CreateJourneyResponse'

Fields

httpStatus :: Int

The response's http status code.

journeyResponse :: JourneyResponse

newCreateJourneyResponse

Arguments

:: Int
$sel:httpStatus:CreateJourneyResponse'

-> JourneyResponse
$sel:journeyResponse:CreateJourneyResponse'

-> CreateJourneyResponse

Create a value of CreateJourneyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateJourneyResponse', createJourneyResponse_httpStatus - The response's http status code.

$sel:journeyResponse:CreateJourneyResponse', createJourneyResponse_journeyResponse - Undocumented member.

Response Lenses

createJourneyResponse_httpStatus :: Lens' CreateJourneyResponse Int

The response's http status code.

createJourneyResponse_journeyResponse :: Lens' CreateJourneyResponse JourneyResponse

Undocumented member.