Network.AWS.Pinpoint.CreateSmsTemplate

Description

Creates a message template for messages that are sent through the SMS channel.

Creating a Request

data CreateSmsTemplate

See: newCreateSmsTemplate smart constructor.

Constructors

CreateSmsTemplate'

Fields

templateName :: Text

The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

sMSTemplateRequest :: SMSTemplateRequest

newCreateSmsTemplate

Arguments

:: Text
$sel:templateName:CreateSmsTemplate'

-> SMSTemplateRequest
$sel:sMSTemplateRequest:CreateSmsTemplate'

-> CreateSmsTemplate

Create a value of CreateSmsTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:CreateSmsTemplate', createSmsTemplate_templateName - The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

$sel:sMSTemplateRequest:CreateSmsTemplate', createSmsTemplate_sMSTemplateRequest - Undocumented member.

Request Lenses

createSmsTemplate_templateName :: Lens' CreateSmsTemplate Text

The name of the message template. A template name must start with an alphanumeric character and can contain a maximum of 128 characters. The characters can be alphanumeric characters, underscores (_), or hyphens (-). Template names are case sensitive.

createSmsTemplate_sMSTemplateRequest :: Lens' CreateSmsTemplate SMSTemplateRequest

Undocumented member.

Destructuring the Response

data CreateSmsTemplateResponse

See: newCreateSmsTemplateResponse smart constructor.

Constructors

CreateSmsTemplateResponse'

Fields

httpStatus :: Int

The response's http status code.

createTemplateMessageBody :: CreateTemplateMessageBody

newCreateSmsTemplateResponse

Arguments

:: Int
$sel:httpStatus:CreateSmsTemplateResponse'

-> CreateTemplateMessageBody
$sel:createTemplateMessageBody:CreateSmsTemplateResponse'

-> CreateSmsTemplateResponse

Create a value of CreateSmsTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateSmsTemplateResponse', createSmsTemplateResponse_httpStatus - The response's http status code.

$sel:createTemplateMessageBody:CreateSmsTemplateResponse', createSmsTemplateResponse_createTemplateMessageBody - Undocumented member.

Response Lenses

createSmsTemplateResponse_httpStatus :: Lens' CreateSmsTemplateResponse Int

The response's http status code.

createSmsTemplateResponse_createTemplateMessageBody :: Lens' CreateSmsTemplateResponse CreateTemplateMessageBody

Undocumented member.