Network.AWS.Pinpoint.DeleteApnsChannel

Description

Disables the APNs channel for an application and deletes any existing settings for the channel.

Creating a Request

data DeleteApnsChannel

See: newDeleteApnsChannel smart constructor.

Constructors

DeleteApnsChannel'

Fields

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

newDeleteApnsChannel

Arguments

:: Text
$sel:applicationId:DeleteApnsChannel'

-> DeleteApnsChannel

Create a value of DeleteApnsChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:DeleteApnsChannel', deleteApnsChannel_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Request Lenses

deleteApnsChannel_applicationId :: Lens' DeleteApnsChannel Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Destructuring the Response

data DeleteApnsChannelResponse

See: newDeleteApnsChannelResponse smart constructor.

Constructors

DeleteApnsChannelResponse'

Fields

httpStatus :: Int

The response's http status code.

aPNSChannelResponse :: APNSChannelResponse

newDeleteApnsChannelResponse

Arguments

:: Int
$sel:httpStatus:DeleteApnsChannelResponse'

-> APNSChannelResponse
$sel:aPNSChannelResponse:DeleteApnsChannelResponse'

-> DeleteApnsChannelResponse

Create a value of DeleteApnsChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteApnsChannelResponse', deleteApnsChannelResponse_httpStatus - The response's http status code.

$sel:aPNSChannelResponse:DeleteApnsChannelResponse', deleteApnsChannelResponse_aPNSChannelResponse - Undocumented member.

Response Lenses

deleteApnsChannelResponse_httpStatus :: Lens' DeleteApnsChannelResponse Int

The response's http status code.

deleteApnsChannelResponse_aPNSChannelResponse :: Lens' DeleteApnsChannelResponse APNSChannelResponse

Undocumented member.