Network.AWS.Pinpoint.DeleteEmailChannel

Description

Disables the email channel for an application and deletes any existing settings for the channel.

Creating a Request

data DeleteEmailChannel

See: newDeleteEmailChannel smart constructor.

Constructors

DeleteEmailChannel'

Fields

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

newDeleteEmailChannel

Arguments

:: Text
$sel:applicationId:DeleteEmailChannel'

-> DeleteEmailChannel

Create a value of DeleteEmailChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:DeleteEmailChannel', deleteEmailChannel_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Request Lenses

deleteEmailChannel_applicationId :: Lens' DeleteEmailChannel Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Destructuring the Response

data DeleteEmailChannelResponse

See: newDeleteEmailChannelResponse smart constructor.

Constructors

DeleteEmailChannelResponse'

Fields

httpStatus :: Int

The response's http status code.

emailChannelResponse :: EmailChannelResponse

newDeleteEmailChannelResponse

Arguments

:: Int
$sel:httpStatus:DeleteEmailChannelResponse'

-> EmailChannelResponse
$sel:emailChannelResponse:DeleteEmailChannelResponse'

-> DeleteEmailChannelResponse

Create a value of DeleteEmailChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteEmailChannelResponse', deleteEmailChannelResponse_httpStatus - The response's http status code.

$sel:emailChannelResponse:DeleteEmailChannelResponse', deleteEmailChannelResponse_emailChannelResponse - Undocumented member.

Response Lenses

deleteEmailChannelResponse_httpStatus :: Lens' DeleteEmailChannelResponse Int

The response's http status code.

deleteEmailChannelResponse_emailChannelResponse :: Lens' DeleteEmailChannelResponse EmailChannelResponse

Undocumented member.