Network.AWS.Pinpoint.DeleteEventStream

Description

Deletes the event stream for an application.

Creating a Request

data DeleteEventStream

See: newDeleteEventStream smart constructor.

Constructors

DeleteEventStream'

Fields

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

newDeleteEventStream

Arguments

:: Text
$sel:applicationId:DeleteEventStream'

-> DeleteEventStream

Create a value of DeleteEventStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:DeleteEventStream', deleteEventStream_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Request Lenses

deleteEventStream_applicationId :: Lens' DeleteEventStream Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Destructuring the Response

data DeleteEventStreamResponse

See: newDeleteEventStreamResponse smart constructor.

Constructors

DeleteEventStreamResponse'

Fields

httpStatus :: Int

The response's http status code.

eventStream :: EventStream

newDeleteEventStreamResponse

Arguments

:: Int
$sel:httpStatus:DeleteEventStreamResponse'

-> EventStream
$sel:eventStream:DeleteEventStreamResponse'

-> DeleteEventStreamResponse

Create a value of DeleteEventStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteEventStreamResponse', deleteEventStreamResponse_httpStatus - The response's http status code.

$sel:eventStream:DeleteEventStreamResponse', deleteEventStreamResponse_eventStream - Undocumented member.

Response Lenses

deleteEventStreamResponse_httpStatus :: Lens' DeleteEventStreamResponse Int

The response's http status code.

deleteEventStreamResponse_eventStream :: Lens' DeleteEventStreamResponse EventStream

Undocumented member.