Network.AWS.Pinpoint.DeleteSmsChannel

Description

Disables the SMS channel for an application and deletes any existing settings for the channel.

Creating a Request

data DeleteSmsChannel

See: newDeleteSmsChannel smart constructor.

Constructors

DeleteSmsChannel'

Fields

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

newDeleteSmsChannel

Arguments

:: Text
$sel:applicationId:DeleteSmsChannel'

-> DeleteSmsChannel

Create a value of DeleteSmsChannel with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:DeleteSmsChannel', deleteSmsChannel_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Request Lenses

deleteSmsChannel_applicationId :: Lens' DeleteSmsChannel Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Destructuring the Response

data DeleteSmsChannelResponse

See: newDeleteSmsChannelResponse smart constructor.

Constructors

DeleteSmsChannelResponse'

Fields

httpStatus :: Int

The response's http status code.

sMSChannelResponse :: SMSChannelResponse

newDeleteSmsChannelResponse

Arguments

:: Int
$sel:httpStatus:DeleteSmsChannelResponse'

-> SMSChannelResponse
$sel:sMSChannelResponse:DeleteSmsChannelResponse'

-> DeleteSmsChannelResponse

Create a value of DeleteSmsChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteSmsChannelResponse', deleteSmsChannelResponse_httpStatus - The response's http status code.

$sel:sMSChannelResponse:DeleteSmsChannelResponse', deleteSmsChannelResponse_sMSChannelResponse - Undocumented member.

Response Lenses

deleteSmsChannelResponse_httpStatus :: Lens' DeleteSmsChannelResponse Int

The response's http status code.

deleteSmsChannelResponse_sMSChannelResponse :: Lens' DeleteSmsChannelResponse SMSChannelResponse

Undocumented member.