Network.AWS.Pinpoint.GetApplicationDateRangeKpi

Description

Retrieves (queries) pre-aggregated data for a standard metric that applies to an application.

Creating a Request

data GetApplicationDateRangeKpi

See: newGetApplicationDateRangeKpi smart constructor.

Constructors

GetApplicationDateRangeKpi'

Fields

nextToken :: Maybe Text

The string that specifies which page of results to return in a paginated response. This parameter is not supported for application, campaign, and journey metrics.

pageSize :: Maybe Text

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

startTime :: Maybe POSIX

The first date and time to retrieve data for, as part of an inclusive date range that filters the query results. This value should be in extended ISO 8601 format and use Coordinated Universal Time (UTC), for example: 2019-07-19T20:00:00Z for 8:00 PM UTC July 19, 2019. This value should also be fewer than 90 days from the current day.

endTime :: Maybe POSIX

The last date and time to retrieve data for, as part of an inclusive date range that filters the query results. This value should be in extended ISO 8601 format and use Coordinated Universal Time (UTC), for example: 2019-07-26T20:00:00Z for 8:00 PM UTC July 26, 2019.

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

kpiName :: Text

The name of the metric, also referred to as a /key performance indicator (KPI)/, to retrieve data for. This value describes the associated metric and consists of two or more terms, which are comprised of lowercase alphanumeric characters, separated by a hyphen. Examples are email-open-rate and successful-delivery-rate. For a list of valid values, see the Amazon Pinpoint Developer Guide.

newGetApplicationDateRangeKpi

Arguments

:: Text
$sel:applicationId:GetApplicationDateRangeKpi'

-> Text
$sel:kpiName:GetApplicationDateRangeKpi'

-> GetApplicationDateRangeKpi

Create a value of GetApplicationDateRangeKpi with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetApplicationDateRangeKpi', getApplicationDateRangeKpi_nextToken - The string that specifies which page of results to return in a paginated response. This parameter is not supported for application, campaign, and journey metrics.

$sel:pageSize:GetApplicationDateRangeKpi', getApplicationDateRangeKpi_pageSize - The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

$sel:startTime:GetApplicationDateRangeKpi', getApplicationDateRangeKpi_startTime - The first date and time to retrieve data for, as part of an inclusive date range that filters the query results. This value should be in extended ISO 8601 format and use Coordinated Universal Time (UTC), for example: 2019-07-19T20:00:00Z for 8:00 PM UTC July 19, 2019. This value should also be fewer than 90 days from the current day.

$sel:endTime:GetApplicationDateRangeKpi', getApplicationDateRangeKpi_endTime - The last date and time to retrieve data for, as part of an inclusive date range that filters the query results. This value should be in extended ISO 8601 format and use Coordinated Universal Time (UTC), for example: 2019-07-26T20:00:00Z for 8:00 PM UTC July 26, 2019.

$sel:applicationId:GetApplicationDateRangeKpi', getApplicationDateRangeKpi_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

$sel:kpiName:GetApplicationDateRangeKpi', getApplicationDateRangeKpi_kpiName - The name of the metric, also referred to as a /key performance indicator (KPI)/, to retrieve data for. This value describes the associated metric and consists of two or more terms, which are comprised of lowercase alphanumeric characters, separated by a hyphen. Examples are email-open-rate and successful-delivery-rate. For a list of valid values, see the Amazon Pinpoint Developer Guide.

Request Lenses

getApplicationDateRangeKpi_nextToken :: Lens' GetApplicationDateRangeKpi (Maybe Text)

The string that specifies which page of results to return in a paginated response. This parameter is not supported for application, campaign, and journey metrics.

getApplicationDateRangeKpi_pageSize :: Lens' GetApplicationDateRangeKpi (Maybe Text)

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

getApplicationDateRangeKpi_startTime :: Lens' GetApplicationDateRangeKpi (Maybe UTCTime)

The first date and time to retrieve data for, as part of an inclusive date range that filters the query results. This value should be in extended ISO 8601 format and use Coordinated Universal Time (UTC), for example: 2019-07-19T20:00:00Z for 8:00 PM UTC July 19, 2019. This value should also be fewer than 90 days from the current day.

getApplicationDateRangeKpi_endTime :: Lens' GetApplicationDateRangeKpi (Maybe UTCTime)

The last date and time to retrieve data for, as part of an inclusive date range that filters the query results. This value should be in extended ISO 8601 format and use Coordinated Universal Time (UTC), for example: 2019-07-26T20:00:00Z for 8:00 PM UTC July 26, 2019.

getApplicationDateRangeKpi_applicationId :: Lens' GetApplicationDateRangeKpi Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

getApplicationDateRangeKpi_kpiName :: Lens' GetApplicationDateRangeKpi Text

The name of the metric, also referred to as a /key performance indicator (KPI)/, to retrieve data for. This value describes the associated metric and consists of two or more terms, which are comprised of lowercase alphanumeric characters, separated by a hyphen. Examples are email-open-rate and successful-delivery-rate. For a list of valid values, see the Amazon Pinpoint Developer Guide.

Destructuring the Response

data GetApplicationDateRangeKpiResponse

See: newGetApplicationDateRangeKpiResponse smart constructor.

Constructors

GetApplicationDateRangeKpiResponse'

Fields

httpStatus :: Int

The response's http status code.

applicationDateRangeKpiResponse :: ApplicationDateRangeKpiResponse

newGetApplicationDateRangeKpiResponse

Arguments

:: Int
$sel:httpStatus:GetApplicationDateRangeKpiResponse'

-> ApplicationDateRangeKpiResponse
$sel:applicationDateRangeKpiResponse:GetApplicationDateRangeKpiResponse'

-> GetApplicationDateRangeKpiResponse

Create a value of GetApplicationDateRangeKpiResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetApplicationDateRangeKpiResponse', getApplicationDateRangeKpiResponse_httpStatus - The response's http status code.

$sel:applicationDateRangeKpiResponse:GetApplicationDateRangeKpiResponse', getApplicationDateRangeKpiResponse_applicationDateRangeKpiResponse - Undocumented member.

Response Lenses

getApplicationDateRangeKpiResponse_httpStatus :: Lens' GetApplicationDateRangeKpiResponse Int

The response's http status code.

getApplicationDateRangeKpiResponse_applicationDateRangeKpiResponse :: Lens' GetApplicationDateRangeKpiResponse ApplicationDateRangeKpiResponse

Undocumented member.