Network.AWS.Pinpoint.GetCampaignDateRangeKpi

Description

Retrieves (queries) pre-aggregated data for a standard metric that applies to a campaign.

Creating a Request

data GetCampaignDateRangeKpi

See: newGetCampaignDateRangeKpi smart constructor.

Constructors

GetCampaignDateRangeKpi'

Fields

nextToken :: Maybe Text

The string that specifies which page of results to return in a paginated response. This parameter is not supported for application, campaign, and journey metrics.

pageSize :: Maybe Text

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

startTime :: Maybe POSIX

The first date and time to retrieve data for, as part of an inclusive date range that filters the query results. This value should be in extended ISO 8601 format and use Coordinated Universal Time (UTC), for example: 2019-07-19T20:00:00Z for 8:00 PM UTC July 19, 2019. This value should also be fewer than 90 days from the current day.

endTime :: Maybe POSIX

The last date and time to retrieve data for, as part of an inclusive date range that filters the query results. This value should be in extended ISO 8601 format and use Coordinated Universal Time (UTC), for example: 2019-07-26T20:00:00Z for 8:00 PM UTC July 26, 2019.

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

kpiName :: Text

The name of the metric, also referred to as a /key performance indicator (KPI)/, to retrieve data for. This value describes the associated metric and consists of two or more terms, which are comprised of lowercase alphanumeric characters, separated by a hyphen. Examples are email-open-rate and successful-delivery-rate. For a list of valid values, see the Amazon Pinpoint Developer Guide.

campaignId :: Text

The unique identifier for the campaign.

newGetCampaignDateRangeKpi

Arguments

:: Text
$sel:applicationId:GetCampaignDateRangeKpi'

-> Text
$sel:kpiName:GetCampaignDateRangeKpi'

-> Text
$sel:campaignId:GetCampaignDateRangeKpi'

-> GetCampaignDateRangeKpi

Create a value of GetCampaignDateRangeKpi with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetCampaignDateRangeKpi', getCampaignDateRangeKpi_nextToken - The string that specifies which page of results to return in a paginated response. This parameter is not supported for application, campaign, and journey metrics.

$sel:pageSize:GetCampaignDateRangeKpi', getCampaignDateRangeKpi_pageSize - The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

$sel:startTime:GetCampaignDateRangeKpi', getCampaignDateRangeKpi_startTime - The first date and time to retrieve data for, as part of an inclusive date range that filters the query results. This value should be in extended ISO 8601 format and use Coordinated Universal Time (UTC), for example: 2019-07-19T20:00:00Z for 8:00 PM UTC July 19, 2019. This value should also be fewer than 90 days from the current day.

$sel:endTime:GetCampaignDateRangeKpi', getCampaignDateRangeKpi_endTime - The last date and time to retrieve data for, as part of an inclusive date range that filters the query results. This value should be in extended ISO 8601 format and use Coordinated Universal Time (UTC), for example: 2019-07-26T20:00:00Z for 8:00 PM UTC July 26, 2019.

$sel:applicationId:GetCampaignDateRangeKpi', getCampaignDateRangeKpi_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

$sel:kpiName:GetCampaignDateRangeKpi', getCampaignDateRangeKpi_kpiName - The name of the metric, also referred to as a /key performance indicator (KPI)/, to retrieve data for. This value describes the associated metric and consists of two or more terms, which are comprised of lowercase alphanumeric characters, separated by a hyphen. Examples are email-open-rate and successful-delivery-rate. For a list of valid values, see the Amazon Pinpoint Developer Guide.

$sel:campaignId:GetCampaignDateRangeKpi', getCampaignDateRangeKpi_campaignId - The unique identifier for the campaign.

Request Lenses

getCampaignDateRangeKpi_nextToken :: Lens' GetCampaignDateRangeKpi (Maybe Text)

The string that specifies which page of results to return in a paginated response. This parameter is not supported for application, campaign, and journey metrics.

getCampaignDateRangeKpi_pageSize :: Lens' GetCampaignDateRangeKpi (Maybe Text)

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

getCampaignDateRangeKpi_startTime :: Lens' GetCampaignDateRangeKpi (Maybe UTCTime)

The first date and time to retrieve data for, as part of an inclusive date range that filters the query results. This value should be in extended ISO 8601 format and use Coordinated Universal Time (UTC), for example: 2019-07-19T20:00:00Z for 8:00 PM UTC July 19, 2019. This value should also be fewer than 90 days from the current day.

getCampaignDateRangeKpi_endTime :: Lens' GetCampaignDateRangeKpi (Maybe UTCTime)

The last date and time to retrieve data for, as part of an inclusive date range that filters the query results. This value should be in extended ISO 8601 format and use Coordinated Universal Time (UTC), for example: 2019-07-26T20:00:00Z for 8:00 PM UTC July 26, 2019.

getCampaignDateRangeKpi_applicationId :: Lens' GetCampaignDateRangeKpi Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

getCampaignDateRangeKpi_kpiName :: Lens' GetCampaignDateRangeKpi Text

The name of the metric, also referred to as a /key performance indicator (KPI)/, to retrieve data for. This value describes the associated metric and consists of two or more terms, which are comprised of lowercase alphanumeric characters, separated by a hyphen. Examples are email-open-rate and successful-delivery-rate. For a list of valid values, see the Amazon Pinpoint Developer Guide.

getCampaignDateRangeKpi_campaignId :: Lens' GetCampaignDateRangeKpi Text

The unique identifier for the campaign.

Destructuring the Response

data GetCampaignDateRangeKpiResponse

See: newGetCampaignDateRangeKpiResponse smart constructor.

Constructors

GetCampaignDateRangeKpiResponse'

Fields

httpStatus :: Int

The response's http status code.

campaignDateRangeKpiResponse :: CampaignDateRangeKpiResponse

newGetCampaignDateRangeKpiResponse

Arguments

:: Int
$sel:httpStatus:GetCampaignDateRangeKpiResponse'

-> CampaignDateRangeKpiResponse
$sel:campaignDateRangeKpiResponse:GetCampaignDateRangeKpiResponse'

-> GetCampaignDateRangeKpiResponse

Create a value of GetCampaignDateRangeKpiResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetCampaignDateRangeKpiResponse', getCampaignDateRangeKpiResponse_httpStatus - The response's http status code.

$sel:campaignDateRangeKpiResponse:GetCampaignDateRangeKpiResponse', getCampaignDateRangeKpiResponse_campaignDateRangeKpiResponse - Undocumented member.

Response Lenses

getCampaignDateRangeKpiResponse_httpStatus :: Lens' GetCampaignDateRangeKpiResponse Int

The response's http status code.

getCampaignDateRangeKpiResponse_campaignDateRangeKpiResponse :: Lens' GetCampaignDateRangeKpiResponse CampaignDateRangeKpiResponse

Undocumented member.