Network.AWS.Pinpoint.GetEventStream

Description

Retrieves information about the event stream settings for an application.

Creating a Request

data GetEventStream

See: newGetEventStream smart constructor.

Constructors

GetEventStream'

Fields

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

newGetEventStream

Arguments

:: Text
$sel:applicationId:GetEventStream'

-> GetEventStream

Create a value of GetEventStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:GetEventStream', getEventStream_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Request Lenses

getEventStream_applicationId :: Lens' GetEventStream Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Destructuring the Response

data GetEventStreamResponse

See: newGetEventStreamResponse smart constructor.

Constructors

GetEventStreamResponse'

Fields

httpStatus :: Int

The response's http status code.

eventStream :: EventStream

newGetEventStreamResponse

Arguments

:: Int
$sel:httpStatus:GetEventStreamResponse'

-> EventStream
$sel:eventStream:GetEventStreamResponse'

-> GetEventStreamResponse

Create a value of GetEventStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetEventStreamResponse', getEventStreamResponse_httpStatus - The response's http status code.

$sel:eventStream:GetEventStreamResponse', getEventStreamResponse_eventStream - Undocumented member.

Response Lenses

getEventStreamResponse_httpStatus :: Lens' GetEventStreamResponse Int

The response's http status code.

getEventStreamResponse_eventStream :: Lens' GetEventStreamResponse EventStream

Undocumented member.