Network.AWS.Pinpoint.GetExportJobs

Description

Retrieves information about the status and settings of all the export jobs for an application.

Creating a Request

data GetExportJobs

See: newGetExportJobs smart constructor.

Constructors

GetExportJobs'

Fields

pageSize :: Maybe Text

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

token :: Maybe Text

The NextToken string that specifies which page of results to return in a paginated response.

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

newGetExportJobs

Arguments

:: Text
$sel:applicationId:GetExportJobs'

-> GetExportJobs

Create a value of GetExportJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pageSize:GetExportJobs', getExportJobs_pageSize - The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

$sel:token:GetExportJobs', getExportJobs_token - The NextToken string that specifies which page of results to return in a paginated response.

$sel:applicationId:GetExportJobs', getExportJobs_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Request Lenses

getExportJobs_pageSize :: Lens' GetExportJobs (Maybe Text)

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

getExportJobs_token :: Lens' GetExportJobs (Maybe Text)

The NextToken string that specifies which page of results to return in a paginated response.

getExportJobs_applicationId :: Lens' GetExportJobs Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Destructuring the Response

data GetExportJobsResponse

See: newGetExportJobsResponse smart constructor.

Constructors

GetExportJobsResponse'

Fields

httpStatus :: Int

The response's http status code.

exportJobsResponse :: ExportJobsResponse

newGetExportJobsResponse

Arguments

:: Int
$sel:httpStatus:GetExportJobsResponse'

-> ExportJobsResponse
$sel:exportJobsResponse:GetExportJobsResponse'

-> GetExportJobsResponse

Create a value of GetExportJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetExportJobsResponse', getExportJobsResponse_httpStatus - The response's http status code.

$sel:exportJobsResponse:GetExportJobsResponse', getExportJobsResponse_exportJobsResponse - Undocumented member.

Response Lenses

getExportJobsResponse_httpStatus :: Lens' GetExportJobsResponse Int

The response's http status code.

getExportJobsResponse_exportJobsResponse :: Lens' GetExportJobsResponse ExportJobsResponse

Undocumented member.