Network.AWS.Pinpoint.GetInAppMessages

Description

Retrieves the in-app messages targeted for the provided endpoint ID.

Creating a Request

data GetInAppMessages

See: newGetInAppMessages smart constructor.

Constructors

GetInAppMessages'

Fields

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

endpointId :: Text

The unique identifier for the endpoint.

newGetInAppMessages

Arguments

:: Text
$sel:applicationId:GetInAppMessages'

-> Text
$sel:endpointId:GetInAppMessages'

-> GetInAppMessages

Create a value of GetInAppMessages with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:GetInAppMessages', getInAppMessages_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

$sel:endpointId:GetInAppMessages', getInAppMessages_endpointId - The unique identifier for the endpoint.

Request Lenses

getInAppMessages_applicationId :: Lens' GetInAppMessages Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

getInAppMessages_endpointId :: Lens' GetInAppMessages Text

The unique identifier for the endpoint.

Destructuring the Response

data GetInAppMessagesResponse

See: newGetInAppMessagesResponse smart constructor.

Constructors

GetInAppMessagesResponse'

Fields

httpStatus :: Int

The response's http status code.

inAppMessagesResponse :: InAppMessagesResponse

newGetInAppMessagesResponse

Arguments

:: Int
$sel:httpStatus:GetInAppMessagesResponse'

-> InAppMessagesResponse
$sel:inAppMessagesResponse:GetInAppMessagesResponse'

-> GetInAppMessagesResponse

Create a value of GetInAppMessagesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetInAppMessagesResponse', getInAppMessagesResponse_httpStatus - The response's http status code.

$sel:inAppMessagesResponse:GetInAppMessagesResponse', getInAppMessagesResponse_inAppMessagesResponse - Undocumented member.

Response Lenses

getInAppMessagesResponse_httpStatus :: Lens' GetInAppMessagesResponse Int

The response's http status code.

getInAppMessagesResponse_inAppMessagesResponse :: Lens' GetInAppMessagesResponse InAppMessagesResponse

Undocumented member.