Network.AWS.Pinpoint.GetJourneyExecutionActivityMetrics

Description

Retrieves (queries) pre-aggregated data for a standard execution metric that applies to a journey activity.

Creating a Request

data GetJourneyExecutionActivityMetrics

See: newGetJourneyExecutionActivityMetrics smart constructor.

Constructors

GetJourneyExecutionActivityMetrics'

Fields

nextToken :: Maybe Text

The string that specifies which page of results to return in a paginated response. This parameter is not supported for application, campaign, and journey metrics.

pageSize :: Maybe Text

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

journeyActivityId :: Text

The unique identifier for the journey activity.

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

journeyId :: Text

The unique identifier for the journey.

newGetJourneyExecutionActivityMetrics

Arguments

:: Text
$sel:journeyActivityId:GetJourneyExecutionActivityMetrics'

-> Text
$sel:applicationId:GetJourneyExecutionActivityMetrics'

-> Text
$sel:journeyId:GetJourneyExecutionActivityMetrics'

-> GetJourneyExecutionActivityMetrics

Create a value of GetJourneyExecutionActivityMetrics with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetJourneyExecutionActivityMetrics', getJourneyExecutionActivityMetrics_nextToken - The string that specifies which page of results to return in a paginated response. This parameter is not supported for application, campaign, and journey metrics.

$sel:pageSize:GetJourneyExecutionActivityMetrics', getJourneyExecutionActivityMetrics_pageSize - The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

$sel:journeyActivityId:GetJourneyExecutionActivityMetrics', getJourneyExecutionActivityMetrics_journeyActivityId - The unique identifier for the journey activity.

$sel:applicationId:GetJourneyExecutionActivityMetrics', getJourneyExecutionActivityMetrics_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

$sel:journeyId:GetJourneyExecutionActivityMetrics', getJourneyExecutionActivityMetrics_journeyId - The unique identifier for the journey.

Request Lenses

getJourneyExecutionActivityMetrics_nextToken :: Lens' GetJourneyExecutionActivityMetrics (Maybe Text)

The string that specifies which page of results to return in a paginated response. This parameter is not supported for application, campaign, and journey metrics.

getJourneyExecutionActivityMetrics_pageSize :: Lens' GetJourneyExecutionActivityMetrics (Maybe Text)

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

getJourneyExecutionActivityMetrics_journeyActivityId :: Lens' GetJourneyExecutionActivityMetrics Text

The unique identifier for the journey activity.

getJourneyExecutionActivityMetrics_applicationId :: Lens' GetJourneyExecutionActivityMetrics Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

getJourneyExecutionActivityMetrics_journeyId :: Lens' GetJourneyExecutionActivityMetrics Text

The unique identifier for the journey.

Destructuring the Response

data GetJourneyExecutionActivityMetricsResponse

See: newGetJourneyExecutionActivityMetricsResponse smart constructor.

Constructors

GetJourneyExecutionActivityMetricsResponse'

Fields

httpStatus :: Int

The response's http status code.

journeyExecutionActivityMetricsResponse :: JourneyExecutionActivityMetricsResponse

newGetJourneyExecutionActivityMetricsResponse

Arguments

:: Int
$sel:httpStatus:GetJourneyExecutionActivityMetricsResponse'

-> JourneyExecutionActivityMetricsResponse
$sel:journeyExecutionActivityMetricsResponse:GetJourneyExecutionActivityMetricsResponse'

-> GetJourneyExecutionActivityMetricsResponse

Create a value of GetJourneyExecutionActivityMetricsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetJourneyExecutionActivityMetricsResponse', getJourneyExecutionActivityMetricsResponse_httpStatus - The response's http status code.

$sel:journeyExecutionActivityMetricsResponse:GetJourneyExecutionActivityMetricsResponse', getJourneyExecutionActivityMetricsResponse_journeyExecutionActivityMetricsResponse - Undocumented member.

Response Lenses

getJourneyExecutionActivityMetricsResponse_httpStatus :: Lens' GetJourneyExecutionActivityMetricsResponse Int

The response's http status code.

getJourneyExecutionActivityMetricsResponse_journeyExecutionActivityMetricsResponse :: Lens' GetJourneyExecutionActivityMetricsResponse JourneyExecutionActivityMetricsResponse

Undocumented member.