Network.AWS.Pinpoint.GetRecommenderConfiguration

Description

Retrieves information about an Amazon Pinpoint configuration for a recommender model.

Creating a Request

data GetRecommenderConfiguration

See: newGetRecommenderConfiguration smart constructor.

Constructors

GetRecommenderConfiguration'

Fields

recommenderId :: Text

The unique identifier for the recommender model configuration. This identifier is displayed as the Recommender ID on the Amazon Pinpoint console.

newGetRecommenderConfiguration

Arguments

:: Text
$sel:recommenderId:GetRecommenderConfiguration'

-> GetRecommenderConfiguration

Create a value of GetRecommenderConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recommenderId:GetRecommenderConfiguration', getRecommenderConfiguration_recommenderId - The unique identifier for the recommender model configuration. This identifier is displayed as the Recommender ID on the Amazon Pinpoint console.

Request Lenses

getRecommenderConfiguration_recommenderId :: Lens' GetRecommenderConfiguration Text

The unique identifier for the recommender model configuration. This identifier is displayed as the Recommender ID on the Amazon Pinpoint console.

Destructuring the Response

data GetRecommenderConfigurationResponse

See: newGetRecommenderConfigurationResponse smart constructor.

Constructors

GetRecommenderConfigurationResponse'

Fields

httpStatus :: Int

The response's http status code.

recommenderConfigurationResponse :: RecommenderConfigurationResponse

newGetRecommenderConfigurationResponse

Arguments

:: Int
$sel:httpStatus:GetRecommenderConfigurationResponse'

-> RecommenderConfigurationResponse
$sel:recommenderConfigurationResponse:GetRecommenderConfigurationResponse'

-> GetRecommenderConfigurationResponse

Create a value of GetRecommenderConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetRecommenderConfigurationResponse', getRecommenderConfigurationResponse_httpStatus - The response's http status code.

$sel:recommenderConfigurationResponse:GetRecommenderConfigurationResponse', getRecommenderConfigurationResponse_recommenderConfigurationResponse - Undocumented member.

Response Lenses

getRecommenderConfigurationResponse_httpStatus :: Lens' GetRecommenderConfigurationResponse Int

The response's http status code.

getRecommenderConfigurationResponse_recommenderConfigurationResponse :: Lens' GetRecommenderConfigurationResponse RecommenderConfigurationResponse

Undocumented member.