Network.AWS.Pinpoint.GetRecommenderConfigurations

Description

Retrieves information about all the recommender model configurations that are associated with your Amazon Pinpoint account.

Creating a Request

data GetRecommenderConfigurations

See: newGetRecommenderConfigurations smart constructor.

Constructors

GetRecommenderConfigurations'

Fields

token :: Maybe Text

The NextToken string that specifies which page of results to return in a paginated response.

pageSize :: Maybe Text

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

newGetRecommenderConfigurations :: GetRecommenderConfigurations

Create a value of GetRecommenderConfigurations with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:token:GetRecommenderConfigurations', getRecommenderConfigurations_token - The NextToken string that specifies which page of results to return in a paginated response.

$sel:pageSize:GetRecommenderConfigurations', getRecommenderConfigurations_pageSize - The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

Request Lenses

getRecommenderConfigurations_token :: Lens' GetRecommenderConfigurations (Maybe Text)

The NextToken string that specifies which page of results to return in a paginated response.

getRecommenderConfigurations_pageSize :: Lens' GetRecommenderConfigurations (Maybe Text)

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

Destructuring the Response

data GetRecommenderConfigurationsResponse

See: newGetRecommenderConfigurationsResponse smart constructor.

Constructors

GetRecommenderConfigurationsResponse'

Fields

httpStatus :: Int

The response's http status code.

listRecommenderConfigurationsResponse :: ListRecommenderConfigurationsResponse

newGetRecommenderConfigurationsResponse

Arguments

:: Int
$sel:httpStatus:GetRecommenderConfigurationsResponse'

-> ListRecommenderConfigurationsResponse
$sel:listRecommenderConfigurationsResponse:GetRecommenderConfigurationsResponse'

-> GetRecommenderConfigurationsResponse

Create a value of GetRecommenderConfigurationsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetRecommenderConfigurationsResponse', getRecommenderConfigurationsResponse_httpStatus - The response's http status code.

$sel:listRecommenderConfigurationsResponse:GetRecommenderConfigurationsResponse', getRecommenderConfigurationsResponse_listRecommenderConfigurationsResponse - Undocumented member.

Response Lenses

getRecommenderConfigurationsResponse_httpStatus :: Lens' GetRecommenderConfigurationsResponse Int

The response's http status code.

getRecommenderConfigurationsResponse_listRecommenderConfigurationsResponse :: Lens' GetRecommenderConfigurationsResponse ListRecommenderConfigurationsResponse

Undocumented member.