Network.AWS.Pinpoint.GetSegmentExportJobs

Description

Retrieves information about the status and settings of the export jobs for a segment.

Creating a Request

data GetSegmentExportJobs

See: newGetSegmentExportJobs smart constructor.

Constructors

GetSegmentExportJobs'

Fields

token :: Maybe Text

The NextToken string that specifies which page of results to return in a paginated response.

pageSize :: Maybe Text

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

segmentId :: Text

The unique identifier for the segment.

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

newGetSegmentExportJobs

Arguments

:: Text
$sel:segmentId:GetSegmentExportJobs'

-> Text
$sel:applicationId:GetSegmentExportJobs'

-> GetSegmentExportJobs

Create a value of GetSegmentExportJobs with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:token:GetSegmentExportJobs', getSegmentExportJobs_token - The NextToken string that specifies which page of results to return in a paginated response.

$sel:pageSize:GetSegmentExportJobs', getSegmentExportJobs_pageSize - The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

$sel:segmentId:GetSegmentExportJobs', getSegmentExportJobs_segmentId - The unique identifier for the segment.

$sel:applicationId:GetSegmentExportJobs', getSegmentExportJobs_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Request Lenses

getSegmentExportJobs_token :: Lens' GetSegmentExportJobs (Maybe Text)

The NextToken string that specifies which page of results to return in a paginated response.

getSegmentExportJobs_pageSize :: Lens' GetSegmentExportJobs (Maybe Text)

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

getSegmentExportJobs_segmentId :: Lens' GetSegmentExportJobs Text

The unique identifier for the segment.

getSegmentExportJobs_applicationId :: Lens' GetSegmentExportJobs Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Destructuring the Response

data GetSegmentExportJobsResponse

See: newGetSegmentExportJobsResponse smart constructor.

Constructors

GetSegmentExportJobsResponse'

Fields

httpStatus :: Int

The response's http status code.

exportJobsResponse :: ExportJobsResponse

newGetSegmentExportJobsResponse

Arguments

:: Int
$sel:httpStatus:GetSegmentExportJobsResponse'

-> ExportJobsResponse
$sel:exportJobsResponse:GetSegmentExportJobsResponse'

-> GetSegmentExportJobsResponse

Create a value of GetSegmentExportJobsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetSegmentExportJobsResponse', getSegmentExportJobsResponse_httpStatus - The response's http status code.

$sel:exportJobsResponse:GetSegmentExportJobsResponse', getSegmentExportJobsResponse_exportJobsResponse - Undocumented member.

Response Lenses

getSegmentExportJobsResponse_httpStatus :: Lens' GetSegmentExportJobsResponse Int

The response's http status code.

getSegmentExportJobsResponse_exportJobsResponse :: Lens' GetSegmentExportJobsResponse ExportJobsResponse

Undocumented member.