Network.AWS.Pinpoint.GetSegments

Description

Retrieves information about the configuration, dimension, and other settings for all the segments that are associated with an application.

Creating a Request

data GetSegments

See: newGetSegments smart constructor.

Constructors

GetSegments'

Fields

pageSize :: Maybe Text

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

token :: Maybe Text

The NextToken string that specifies which page of results to return in a paginated response.

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

newGetSegments

Arguments

:: Text
$sel:applicationId:GetSegments'

-> GetSegments

Create a value of GetSegments with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:pageSize:GetSegments', getSegments_pageSize - The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

$sel:token:GetSegments', getSegments_token - The NextToken string that specifies which page of results to return in a paginated response.

$sel:applicationId:GetSegments', getSegments_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Request Lenses

getSegments_pageSize :: Lens' GetSegments (Maybe Text)

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

getSegments_token :: Lens' GetSegments (Maybe Text)

The NextToken string that specifies which page of results to return in a paginated response.

getSegments_applicationId :: Lens' GetSegments Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Destructuring the Response

data GetSegmentsResponse

See: newGetSegmentsResponse smart constructor.

Constructors

GetSegmentsResponse'

Fields

httpStatus :: Int

The response's http status code.

segmentsResponse :: SegmentsResponse

newGetSegmentsResponse

Arguments

:: Int
$sel:httpStatus:GetSegmentsResponse'

-> SegmentsResponse
$sel:segmentsResponse:GetSegmentsResponse'

-> GetSegmentsResponse

Create a value of GetSegmentsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetSegmentsResponse', getSegmentsResponse_httpStatus - The response's http status code.

$sel:segmentsResponse:GetSegmentsResponse', getSegmentsResponse_segmentsResponse - Undocumented member.

Response Lenses

getSegmentsResponse_httpStatus :: Lens' GetSegmentsResponse Int

The response's http status code.

getSegmentsResponse_segmentsResponse :: Lens' GetSegmentsResponse SegmentsResponse

Undocumented member.