Network.AWS.Pinpoint.ListJourneys

Description

Retrieves information about the status, configuration, and other settings for all the journeys that are associated with an application.

Creating a Request

data ListJourneys

See: newListJourneys smart constructor.

Constructors

ListJourneys'

Fields

token :: Maybe Text

The NextToken string that specifies which page of results to return in a paginated response.

pageSize :: Maybe Text

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

newListJourneys

Arguments

:: Text
$sel:applicationId:ListJourneys'

-> ListJourneys

Create a value of ListJourneys with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:token:ListJourneys', listJourneys_token - The NextToken string that specifies which page of results to return in a paginated response.

$sel:pageSize:ListJourneys', listJourneys_pageSize - The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

$sel:applicationId:ListJourneys', listJourneys_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Request Lenses

listJourneys_token :: Lens' ListJourneys (Maybe Text)

The NextToken string that specifies which page of results to return in a paginated response.

listJourneys_pageSize :: Lens' ListJourneys (Maybe Text)

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

listJourneys_applicationId :: Lens' ListJourneys Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

Destructuring the Response

data ListJourneysResponse

See: newListJourneysResponse smart constructor.

Constructors

ListJourneysResponse'

Fields

httpStatus :: Int

The response's http status code.

journeysResponse :: JourneysResponse

newListJourneysResponse

Arguments

:: Int
$sel:httpStatus:ListJourneysResponse'

-> JourneysResponse
$sel:journeysResponse:ListJourneysResponse'

-> ListJourneysResponse

Create a value of ListJourneysResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ListJourneysResponse', listJourneysResponse_httpStatus - The response's http status code.

$sel:journeysResponse:ListJourneysResponse', listJourneysResponse_journeysResponse - Undocumented member.

Response Lenses

listJourneysResponse_httpStatus :: Lens' ListJourneysResponse Int

The response's http status code.

listJourneysResponse_journeysResponse :: Lens' ListJourneysResponse JourneysResponse

Undocumented member.