Network.AWS.Pinpoint.ListTemplates

Description

Retrieves information about all the message templates that are associated with your Amazon Pinpoint account.

Creating a Request

data ListTemplates

See: newListTemplates smart constructor.

Constructors

ListTemplates'

Fields

templateType :: Maybe Text

The type of message template to include in the results. Valid values are: EMAIL, PUSH, SMS, and VOICE. To include all types of templates in the results, don't include this parameter in your request.

prefix :: Maybe Text

The substring to match in the names of the message templates to include in the results. If you specify this value, Amazon Pinpoint returns only those templates whose names begin with the value that you specify.

nextToken :: Maybe Text

The string that specifies which page of results to return in a paginated response. This parameter is not supported for application, campaign, and journey metrics.

pageSize :: Maybe Text

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

newListTemplates :: ListTemplates

Create a value of ListTemplates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateType:ListTemplates', listTemplates_templateType - The type of message template to include in the results. Valid values are: EMAIL, PUSH, SMS, and VOICE. To include all types of templates in the results, don't include this parameter in your request.

$sel:prefix:ListTemplates', listTemplates_prefix - The substring to match in the names of the message templates to include in the results. If you specify this value, Amazon Pinpoint returns only those templates whose names begin with the value that you specify.

$sel:nextToken:ListTemplates', listTemplates_nextToken - The string that specifies which page of results to return in a paginated response. This parameter is not supported for application, campaign, and journey metrics.

$sel:pageSize:ListTemplates', listTemplates_pageSize - The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

Request Lenses

listTemplates_templateType :: Lens' ListTemplates (Maybe Text)

The type of message template to include in the results. Valid values are: EMAIL, PUSH, SMS, and VOICE. To include all types of templates in the results, don't include this parameter in your request.

listTemplates_prefix :: Lens' ListTemplates (Maybe Text)

The substring to match in the names of the message templates to include in the results. If you specify this value, Amazon Pinpoint returns only those templates whose names begin with the value that you specify.

listTemplates_nextToken :: Lens' ListTemplates (Maybe Text)

The string that specifies which page of results to return in a paginated response. This parameter is not supported for application, campaign, and journey metrics.

listTemplates_pageSize :: Lens' ListTemplates (Maybe Text)

The maximum number of items to include in each page of a paginated response. This parameter is not supported for application, campaign, and journey metrics.

Destructuring the Response

data ListTemplatesResponse

See: newListTemplatesResponse smart constructor.

Constructors

ListTemplatesResponse'

Fields

httpStatus :: Int

The response's http status code.

templatesResponse :: TemplatesResponse

newListTemplatesResponse

Arguments

:: Int
$sel:httpStatus:ListTemplatesResponse'

-> TemplatesResponse
$sel:templatesResponse:ListTemplatesResponse'

-> ListTemplatesResponse

Create a value of ListTemplatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ListTemplatesResponse', listTemplatesResponse_httpStatus - The response's http status code.

$sel:templatesResponse:ListTemplatesResponse', listTemplatesResponse_templatesResponse - Undocumented member.

Response Lenses

listTemplatesResponse_httpStatus :: Lens' ListTemplatesResponse Int

The response's http status code.

listTemplatesResponse_templatesResponse :: Lens' ListTemplatesResponse TemplatesResponse

Undocumented member.