Network.AWS.Pinpoint.RemoveAttributes

Description

Removes one or more attributes, of the same attribute type, from all the endpoints that are associated with an application.

Creating a Request

data RemoveAttributes

See: newRemoveAttributes smart constructor.

Constructors

RemoveAttributes'

Fields

attributeType :: Text

The type of attribute or attributes to remove. Valid values are:

endpoint-custom-attributes - Custom attributes that describe endpoints, such as the date when an associated user opted in or out of receiving communications from you through a specific type of channel.

endpoint-metric-attributes - Custom metrics that your app reports to Amazon Pinpoint for endpoints, such as the number of app sessions or the number of items left in a cart.

endpoint-user-attributes - Custom attributes that describe users, such as first name, last name, and age.

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

updateAttributesRequest :: UpdateAttributesRequest

newRemoveAttributes

Arguments

:: Text
$sel:attributeType:RemoveAttributes'

-> Text
$sel:applicationId:RemoveAttributes'

-> UpdateAttributesRequest
$sel:updateAttributesRequest:RemoveAttributes'

-> RemoveAttributes

Create a value of RemoveAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeType:RemoveAttributes', removeAttributes_attributeType - The type of attribute or attributes to remove. Valid values are:

endpoint-custom-attributes - Custom attributes that describe endpoints, such as the date when an associated user opted in or out of receiving communications from you through a specific type of channel.

endpoint-metric-attributes - Custom metrics that your app reports to Amazon Pinpoint for endpoints, such as the number of app sessions or the number of items left in a cart.

endpoint-user-attributes - Custom attributes that describe users, such as first name, last name, and age.

$sel:applicationId:RemoveAttributes', removeAttributes_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

$sel:updateAttributesRequest:RemoveAttributes', removeAttributes_updateAttributesRequest - Undocumented member.

Request Lenses

removeAttributes_attributeType :: Lens' RemoveAttributes Text

The type of attribute or attributes to remove. Valid values are:

endpoint-custom-attributes - Custom attributes that describe endpoints, such as the date when an associated user opted in or out of receiving communications from you through a specific type of channel.

endpoint-metric-attributes - Custom metrics that your app reports to Amazon Pinpoint for endpoints, such as the number of app sessions or the number of items left in a cart.

endpoint-user-attributes - Custom attributes that describe users, such as first name, last name, and age.

removeAttributes_applicationId :: Lens' RemoveAttributes Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

removeAttributes_updateAttributesRequest :: Lens' RemoveAttributes UpdateAttributesRequest

Undocumented member.

Destructuring the Response

data RemoveAttributesResponse

See: newRemoveAttributesResponse smart constructor.

Constructors

RemoveAttributesResponse'

Fields

httpStatus :: Int

The response's http status code.

attributesResource :: AttributesResource

newRemoveAttributesResponse

Arguments

:: Int
$sel:httpStatus:RemoveAttributesResponse'

-> AttributesResource
$sel:attributesResource:RemoveAttributesResponse'

-> RemoveAttributesResponse

Create a value of RemoveAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:RemoveAttributesResponse', removeAttributesResponse_httpStatus - The response's http status code.

$sel:attributesResource:RemoveAttributesResponse', removeAttributesResponse_attributesResource - Undocumented member.

Response Lenses

removeAttributesResponse_httpStatus :: Lens' RemoveAttributesResponse Int

The response's http status code.

removeAttributesResponse_attributesResource :: Lens' RemoveAttributesResponse AttributesResource

Undocumented member.