Network.AWS.Pinpoint.Types.APNSPushNotificationTemplate

Description

Documentation

data APNSPushNotificationTemplate

Specifies channel-specific content and settings for a message template that can be used in push notifications that are sent through the APNs (Apple Push Notification service) channel.

See: newAPNSPushNotificationTemplate smart constructor.

Constructors

APNSPushNotificationTemplate'

Fields

body :: Maybe Text

The message body to use in push notifications that are based on the message template.

title :: Maybe Text

The title to use in push notifications that are based on the message template. This title appears above the notification message on a recipient's device.

mediaUrl :: Maybe Text

The URL of an image or video to display in push notifications that are based on the message template.

action :: Maybe Action

The action to occur if a recipient taps a push notification that's based on the message template. Valid values are:

OPEN_APP - Your app opens or it becomes the foreground app if it was sent to the background. This is the default action.

DEEP_LINK - Your app opens and displays a designated user interface in the app. This setting uses the deep-linking features of the iOS platform.

URL - The default mobile browser on the recipient's device opens and loads the web page at a URL that you specify.

url :: Maybe Text

The URL to open in the recipient's default mobile browser, if a recipient taps a push notification that's based on the message template and the value of the Action property is URL.

sound :: Maybe Text

The key for the sound to play when the recipient receives a push notification that's based on the message template. The value for this key is the name of a sound file in your app's main bundle or the Library/Sounds folder in your app's data container. If the sound file can't be found or you specify default for the value, the system plays the default alert sound.

rawContent :: Maybe Text

The raw, JSON-formatted string to use as the payload for push notifications that are based on the message template. If specified, this value overrides all other content for the message template.

newAPNSPushNotificationTemplate :: APNSPushNotificationTemplate

Create a value of APNSPushNotificationTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:body:APNSPushNotificationTemplate', aPNSPushNotificationTemplate_body - The message body to use in push notifications that are based on the message template.

$sel:title:APNSPushNotificationTemplate', aPNSPushNotificationTemplate_title - The title to use in push notifications that are based on the message template. This title appears above the notification message on a recipient's device.

$sel:mediaUrl:APNSPushNotificationTemplate', aPNSPushNotificationTemplate_mediaUrl - The URL of an image or video to display in push notifications that are based on the message template.

$sel:action:APNSPushNotificationTemplate', aPNSPushNotificationTemplate_action - The action to occur if a recipient taps a push notification that's based on the message template. Valid values are:

OPEN_APP - Your app opens or it becomes the foreground app if it was sent to the background. This is the default action.

DEEP_LINK - Your app opens and displays a designated user interface in the app. This setting uses the deep-linking features of the iOS platform.

URL - The default mobile browser on the recipient's device opens and loads the web page at a URL that you specify.

$sel:url:APNSPushNotificationTemplate', aPNSPushNotificationTemplate_url - The URL to open in the recipient's default mobile browser, if a recipient taps a push notification that's based on the message template and the value of the Action property is URL.

$sel:sound:APNSPushNotificationTemplate', aPNSPushNotificationTemplate_sound - The key for the sound to play when the recipient receives a push notification that's based on the message template. The value for this key is the name of a sound file in your app's main bundle or the Library/Sounds folder in your app's data container. If the sound file can't be found or you specify default for the value, the system plays the default alert sound.

$sel:rawContent:APNSPushNotificationTemplate', aPNSPushNotificationTemplate_rawContent - The raw, JSON-formatted string to use as the payload for push notifications that are based on the message template. If specified, this value overrides all other content for the message template.

aPNSPushNotificationTemplate_body :: Lens' APNSPushNotificationTemplate (Maybe Text)

The message body to use in push notifications that are based on the message template.

aPNSPushNotificationTemplate_title :: Lens' APNSPushNotificationTemplate (Maybe Text)

The title to use in push notifications that are based on the message template. This title appears above the notification message on a recipient's device.

aPNSPushNotificationTemplate_mediaUrl :: Lens' APNSPushNotificationTemplate (Maybe Text)

The URL of an image or video to display in push notifications that are based on the message template.

aPNSPushNotificationTemplate_action :: Lens' APNSPushNotificationTemplate (Maybe Action)

The action to occur if a recipient taps a push notification that's based on the message template. Valid values are:

OPEN_APP - Your app opens or it becomes the foreground app if it was sent to the background. This is the default action.

DEEP_LINK - Your app opens and displays a designated user interface in the app. This setting uses the deep-linking features of the iOS platform.

URL - The default mobile browser on the recipient's device opens and loads the web page at a URL that you specify.

aPNSPushNotificationTemplate_url :: Lens' APNSPushNotificationTemplate (Maybe Text)

The URL to open in the recipient's default mobile browser, if a recipient taps a push notification that's based on the message template and the value of the Action property is URL.

aPNSPushNotificationTemplate_sound :: Lens' APNSPushNotificationTemplate (Maybe Text)

The key for the sound to play when the recipient receives a push notification that's based on the message template. The value for this key is the name of a sound file in your app's main bundle or the Library/Sounds folder in your app's data container. If the sound file can't be found or you specify default for the value, the system plays the default alert sound.

aPNSPushNotificationTemplate_rawContent :: Lens' APNSPushNotificationTemplate (Maybe Text)

The raw, JSON-formatted string to use as the payload for push notifications that are based on the message template. If specified, this value overrides all other content for the message template.