Network.AWS.Pinpoint.Types.APNSSandboxChannelRequest

Description

Documentation

data APNSSandboxChannelRequest

Specifies the status and settings of the APNs (Apple Push Notification service) sandbox channel for an application.

See: newAPNSSandboxChannelRequest smart constructor.

Constructors

APNSSandboxChannelRequest'

Fields

defaultAuthenticationMethod :: Maybe Text

The default authentication method that you want Amazon Pinpoint to use when authenticating with the APNs sandbox environment, key or certificate.

tokenKey :: Maybe Text

The authentication key to use for APNs tokens.

bundleId :: Maybe Text

The bundle identifier that's assigned to your iOS app. This identifier is used for APNs tokens.

teamId :: Maybe Text

The identifier that's assigned to your Apple developer account team. This identifier is used for APNs tokens.

enabled :: Maybe Bool

Specifies whether to enable the APNs sandbox channel for the application.

privateKey :: Maybe Text

The private key for the APNs client certificate that you want Amazon Pinpoint to use to communicate with the APNs sandbox environment.

tokenKeyId :: Maybe Text

The key identifier that's assigned to your APNs signing key, if you want Amazon Pinpoint to communicate with the APNs sandbox environment by using APNs tokens.

certificate :: Maybe Text

The APNs client certificate that you received from Apple, if you want Amazon Pinpoint to communicate with the APNs sandbox environment by using an APNs certificate.

newAPNSSandboxChannelRequest :: APNSSandboxChannelRequest

Create a value of APNSSandboxChannelRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defaultAuthenticationMethod:APNSSandboxChannelRequest', aPNSSandboxChannelRequest_defaultAuthenticationMethod - The default authentication method that you want Amazon Pinpoint to use when authenticating with the APNs sandbox environment, key or certificate.

$sel:tokenKey:APNSSandboxChannelRequest', aPNSSandboxChannelRequest_tokenKey - The authentication key to use for APNs tokens.

$sel:bundleId:APNSSandboxChannelRequest', aPNSSandboxChannelRequest_bundleId - The bundle identifier that's assigned to your iOS app. This identifier is used for APNs tokens.

$sel:teamId:APNSSandboxChannelRequest', aPNSSandboxChannelRequest_teamId - The identifier that's assigned to your Apple developer account team. This identifier is used for APNs tokens.

$sel:enabled:APNSSandboxChannelRequest', aPNSSandboxChannelRequest_enabled - Specifies whether to enable the APNs sandbox channel for the application.

$sel:privateKey:APNSSandboxChannelRequest', aPNSSandboxChannelRequest_privateKey - The private key for the APNs client certificate that you want Amazon Pinpoint to use to communicate with the APNs sandbox environment.

$sel:tokenKeyId:APNSSandboxChannelRequest', aPNSSandboxChannelRequest_tokenKeyId - The key identifier that's assigned to your APNs signing key, if you want Amazon Pinpoint to communicate with the APNs sandbox environment by using APNs tokens.

$sel:certificate:APNSSandboxChannelRequest', aPNSSandboxChannelRequest_certificate - The APNs client certificate that you received from Apple, if you want Amazon Pinpoint to communicate with the APNs sandbox environment by using an APNs certificate.

aPNSSandboxChannelRequest_defaultAuthenticationMethod :: Lens' APNSSandboxChannelRequest (Maybe Text)

The default authentication method that you want Amazon Pinpoint to use when authenticating with the APNs sandbox environment, key or certificate.

aPNSSandboxChannelRequest_tokenKey :: Lens' APNSSandboxChannelRequest (Maybe Text)

The authentication key to use for APNs tokens.

aPNSSandboxChannelRequest_bundleId :: Lens' APNSSandboxChannelRequest (Maybe Text)

The bundle identifier that's assigned to your iOS app. This identifier is used for APNs tokens.

aPNSSandboxChannelRequest_teamId :: Lens' APNSSandboxChannelRequest (Maybe Text)

The identifier that's assigned to your Apple developer account team. This identifier is used for APNs tokens.

aPNSSandboxChannelRequest_enabled :: Lens' APNSSandboxChannelRequest (Maybe Bool)

Specifies whether to enable the APNs sandbox channel for the application.

aPNSSandboxChannelRequest_privateKey :: Lens' APNSSandboxChannelRequest (Maybe Text)

The private key for the APNs client certificate that you want Amazon Pinpoint to use to communicate with the APNs sandbox environment.

aPNSSandboxChannelRequest_tokenKeyId :: Lens' APNSSandboxChannelRequest (Maybe Text)

The key identifier that's assigned to your APNs signing key, if you want Amazon Pinpoint to communicate with the APNs sandbox environment by using APNs tokens.

aPNSSandboxChannelRequest_certificate :: Lens' APNSSandboxChannelRequest (Maybe Text)

The APNs client certificate that you received from Apple, if you want Amazon Pinpoint to communicate with the APNs sandbox environment by using an APNs certificate.