Network.AWS.Pinpoint.Types.APNSSandboxChannelResponse

Description

Documentation

data APNSSandboxChannelResponse

Provides information about the status and settings of the APNs (Apple Push Notification service) sandbox channel for an application.

See: newAPNSSandboxChannelResponse smart constructor.

Constructors

APNSSandboxChannelResponse'

Fields

applicationId :: Maybe Text

The unique identifier for the application that the APNs sandbox channel applies to.

lastModifiedDate :: Maybe Text

The date and time when the APNs sandbox channel was last modified.

defaultAuthenticationMethod :: Maybe Text

The default authentication method that Amazon Pinpoint uses to authenticate with the APNs sandbox environment for this channel, key or certificate.

hasCredential :: Maybe Bool

(Not used) This property is retained only for backward compatibility.

hasTokenKey :: Maybe Bool

Specifies whether the APNs sandbox channel is configured to communicate with APNs by using APNs tokens. To provide an authentication key for APNs tokens, set the TokenKey property of the channel.

enabled :: Maybe Bool

Specifies whether the APNs sandbox channel is enabled for the application.

creationDate :: Maybe Text

The date and time when the APNs sandbox channel was enabled.

id :: Maybe Text

(Deprecated) An identifier for the APNs sandbox channel. This property is retained only for backward compatibility.

version :: Maybe Int

The current version of the APNs sandbox channel.

isArchived :: Maybe Bool

Specifies whether the APNs sandbox channel is archived.

lastModifiedBy :: Maybe Text

The user who last modified the APNs sandbox channel.

platform :: Text

The type of messaging or notification platform for the channel. For the APNs sandbox channel, this value is APNS_SANDBOX.

newAPNSSandboxChannelResponse

Arguments

:: Text
$sel:platform:APNSSandboxChannelResponse'

-> APNSSandboxChannelResponse

Create a value of APNSSandboxChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:APNSSandboxChannelResponse', aPNSSandboxChannelResponse_applicationId - The unique identifier for the application that the APNs sandbox channel applies to.

$sel:lastModifiedDate:APNSSandboxChannelResponse', aPNSSandboxChannelResponse_lastModifiedDate - The date and time when the APNs sandbox channel was last modified.

$sel:defaultAuthenticationMethod:APNSSandboxChannelResponse', aPNSSandboxChannelResponse_defaultAuthenticationMethod - The default authentication method that Amazon Pinpoint uses to authenticate with the APNs sandbox environment for this channel, key or certificate.

$sel:hasCredential:APNSSandboxChannelResponse', aPNSSandboxChannelResponse_hasCredential - (Not used) This property is retained only for backward compatibility.

$sel:hasTokenKey:APNSSandboxChannelResponse', aPNSSandboxChannelResponse_hasTokenKey - Specifies whether the APNs sandbox channel is configured to communicate with APNs by using APNs tokens. To provide an authentication key for APNs tokens, set the TokenKey property of the channel.

$sel:enabled:APNSSandboxChannelResponse', aPNSSandboxChannelResponse_enabled - Specifies whether the APNs sandbox channel is enabled for the application.

$sel:creationDate:APNSSandboxChannelResponse', aPNSSandboxChannelResponse_creationDate - The date and time when the APNs sandbox channel was enabled.

$sel:id:APNSSandboxChannelResponse', aPNSSandboxChannelResponse_id - (Deprecated) An identifier for the APNs sandbox channel. This property is retained only for backward compatibility.

$sel:version:APNSSandboxChannelResponse', aPNSSandboxChannelResponse_version - The current version of the APNs sandbox channel.

$sel:isArchived:APNSSandboxChannelResponse', aPNSSandboxChannelResponse_isArchived - Specifies whether the APNs sandbox channel is archived.

$sel:lastModifiedBy:APNSSandboxChannelResponse', aPNSSandboxChannelResponse_lastModifiedBy - The user who last modified the APNs sandbox channel.

$sel:platform:APNSSandboxChannelResponse', aPNSSandboxChannelResponse_platform - The type of messaging or notification platform for the channel. For the APNs sandbox channel, this value is APNS_SANDBOX.

aPNSSandboxChannelResponse_applicationId :: Lens' APNSSandboxChannelResponse (Maybe Text)

The unique identifier for the application that the APNs sandbox channel applies to.

aPNSSandboxChannelResponse_lastModifiedDate :: Lens' APNSSandboxChannelResponse (Maybe Text)

The date and time when the APNs sandbox channel was last modified.

aPNSSandboxChannelResponse_defaultAuthenticationMethod :: Lens' APNSSandboxChannelResponse (Maybe Text)

The default authentication method that Amazon Pinpoint uses to authenticate with the APNs sandbox environment for this channel, key or certificate.

aPNSSandboxChannelResponse_hasCredential :: Lens' APNSSandboxChannelResponse (Maybe Bool)

(Not used) This property is retained only for backward compatibility.

aPNSSandboxChannelResponse_hasTokenKey :: Lens' APNSSandboxChannelResponse (Maybe Bool)

Specifies whether the APNs sandbox channel is configured to communicate with APNs by using APNs tokens. To provide an authentication key for APNs tokens, set the TokenKey property of the channel.

aPNSSandboxChannelResponse_enabled :: Lens' APNSSandboxChannelResponse (Maybe Bool)

Specifies whether the APNs sandbox channel is enabled for the application.

aPNSSandboxChannelResponse_creationDate :: Lens' APNSSandboxChannelResponse (Maybe Text)

The date and time when the APNs sandbox channel was enabled.

aPNSSandboxChannelResponse_id :: Lens' APNSSandboxChannelResponse (Maybe Text)

(Deprecated) An identifier for the APNs sandbox channel. This property is retained only for backward compatibility.

aPNSSandboxChannelResponse_version :: Lens' APNSSandboxChannelResponse (Maybe Int)

The current version of the APNs sandbox channel.

aPNSSandboxChannelResponse_isArchived :: Lens' APNSSandboxChannelResponse (Maybe Bool)

Specifies whether the APNs sandbox channel is archived.

aPNSSandboxChannelResponse_lastModifiedBy :: Lens' APNSSandboxChannelResponse (Maybe Text)

The user who last modified the APNs sandbox channel.

aPNSSandboxChannelResponse_platform :: Lens' APNSSandboxChannelResponse Text

The type of messaging or notification platform for the channel. For the APNs sandbox channel, this value is APNS_SANDBOX.