Network.AWS.Pinpoint.Types.APNSVoipChannelRequest

Description

Documentation

data APNSVoipChannelRequest

Specifies the status and settings of the APNs (Apple Push Notification service) VoIP channel for an application.

See: newAPNSVoipChannelRequest smart constructor.

Constructors

APNSVoipChannelRequest'

Fields

defaultAuthenticationMethod :: Maybe Text

The default authentication method that you want Amazon Pinpoint to use when authenticating with APNs, key or certificate.

tokenKey :: Maybe Text

The authentication key to use for APNs tokens.

bundleId :: Maybe Text

The bundle identifier that's assigned to your iOS app. This identifier is used for APNs tokens.

teamId :: Maybe Text

The identifier that's assigned to your Apple developer account team. This identifier is used for APNs tokens.

enabled :: Maybe Bool

Specifies whether to enable the APNs VoIP channel for the application.

privateKey :: Maybe Text

The private key for the APNs client certificate that you want Amazon Pinpoint to use to communicate with APNs.

tokenKeyId :: Maybe Text

The key identifier that's assigned to your APNs signing key, if you want Amazon Pinpoint to communicate with APNs by using APNs tokens.

certificate :: Maybe Text

The APNs client certificate that you received from Apple, if you want Amazon Pinpoint to communicate with APNs by using an APNs certificate.

newAPNSVoipChannelRequest :: APNSVoipChannelRequest

Create a value of APNSVoipChannelRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defaultAuthenticationMethod:APNSVoipChannelRequest', aPNSVoipChannelRequest_defaultAuthenticationMethod - The default authentication method that you want Amazon Pinpoint to use when authenticating with APNs, key or certificate.

$sel:tokenKey:APNSVoipChannelRequest', aPNSVoipChannelRequest_tokenKey - The authentication key to use for APNs tokens.

$sel:bundleId:APNSVoipChannelRequest', aPNSVoipChannelRequest_bundleId - The bundle identifier that's assigned to your iOS app. This identifier is used for APNs tokens.

$sel:teamId:APNSVoipChannelRequest', aPNSVoipChannelRequest_teamId - The identifier that's assigned to your Apple developer account team. This identifier is used for APNs tokens.

$sel:enabled:APNSVoipChannelRequest', aPNSVoipChannelRequest_enabled - Specifies whether to enable the APNs VoIP channel for the application.

$sel:privateKey:APNSVoipChannelRequest', aPNSVoipChannelRequest_privateKey - The private key for the APNs client certificate that you want Amazon Pinpoint to use to communicate with APNs.

$sel:tokenKeyId:APNSVoipChannelRequest', aPNSVoipChannelRequest_tokenKeyId - The key identifier that's assigned to your APNs signing key, if you want Amazon Pinpoint to communicate with APNs by using APNs tokens.

$sel:certificate:APNSVoipChannelRequest', aPNSVoipChannelRequest_certificate - The APNs client certificate that you received from Apple, if you want Amazon Pinpoint to communicate with APNs by using an APNs certificate.

aPNSVoipChannelRequest_defaultAuthenticationMethod :: Lens' APNSVoipChannelRequest (Maybe Text)

The default authentication method that you want Amazon Pinpoint to use when authenticating with APNs, key or certificate.

aPNSVoipChannelRequest_tokenKey :: Lens' APNSVoipChannelRequest (Maybe Text)

The authentication key to use for APNs tokens.

aPNSVoipChannelRequest_bundleId :: Lens' APNSVoipChannelRequest (Maybe Text)

The bundle identifier that's assigned to your iOS app. This identifier is used for APNs tokens.

aPNSVoipChannelRequest_teamId :: Lens' APNSVoipChannelRequest (Maybe Text)

The identifier that's assigned to your Apple developer account team. This identifier is used for APNs tokens.

aPNSVoipChannelRequest_enabled :: Lens' APNSVoipChannelRequest (Maybe Bool)

Specifies whether to enable the APNs VoIP channel for the application.

aPNSVoipChannelRequest_privateKey :: Lens' APNSVoipChannelRequest (Maybe Text)

The private key for the APNs client certificate that you want Amazon Pinpoint to use to communicate with APNs.

aPNSVoipChannelRequest_tokenKeyId :: Lens' APNSVoipChannelRequest (Maybe Text)

The key identifier that's assigned to your APNs signing key, if you want Amazon Pinpoint to communicate with APNs by using APNs tokens.

aPNSVoipChannelRequest_certificate :: Lens' APNSVoipChannelRequest (Maybe Text)

The APNs client certificate that you received from Apple, if you want Amazon Pinpoint to communicate with APNs by using an APNs certificate.