Network.AWS.Pinpoint.Types.APNSVoipChannelResponse

Description

Documentation

data APNSVoipChannelResponse

Provides information about the status and settings of the APNs (Apple Push Notification service) VoIP channel for an application.

See: newAPNSVoipChannelResponse smart constructor.

Constructors

APNSVoipChannelResponse'

Fields

applicationId :: Maybe Text

The unique identifier for the application that the APNs VoIP channel applies to.

lastModifiedDate :: Maybe Text

The date and time when the APNs VoIP channel was last modified.

defaultAuthenticationMethod :: Maybe Text

The default authentication method that Amazon Pinpoint uses to authenticate with APNs for this channel, key or certificate.

hasCredential :: Maybe Bool

(Not used) This property is retained only for backward compatibility.

hasTokenKey :: Maybe Bool

Specifies whether the APNs VoIP channel is configured to communicate with APNs by using APNs tokens. To provide an authentication key for APNs tokens, set the TokenKey property of the channel.

enabled :: Maybe Bool

Specifies whether the APNs VoIP channel is enabled for the application.

creationDate :: Maybe Text

The date and time when the APNs VoIP channel was enabled.

id :: Maybe Text

(Deprecated) An identifier for the APNs VoIP channel. This property is retained only for backward compatibility.

version :: Maybe Int

The current version of the APNs VoIP channel.

isArchived :: Maybe Bool

Specifies whether the APNs VoIP channel is archived.

lastModifiedBy :: Maybe Text

The user who last modified the APNs VoIP channel.

platform :: Text

The type of messaging or notification platform for the channel. For the APNs VoIP channel, this value is APNS_VOIP.

newAPNSVoipChannelResponse

Arguments

:: Text
$sel:platform:APNSVoipChannelResponse'

-> APNSVoipChannelResponse

Create a value of APNSVoipChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:APNSVoipChannelResponse', aPNSVoipChannelResponse_applicationId - The unique identifier for the application that the APNs VoIP channel applies to.

$sel:lastModifiedDate:APNSVoipChannelResponse', aPNSVoipChannelResponse_lastModifiedDate - The date and time when the APNs VoIP channel was last modified.

$sel:defaultAuthenticationMethod:APNSVoipChannelResponse', aPNSVoipChannelResponse_defaultAuthenticationMethod - The default authentication method that Amazon Pinpoint uses to authenticate with APNs for this channel, key or certificate.

$sel:hasCredential:APNSVoipChannelResponse', aPNSVoipChannelResponse_hasCredential - (Not used) This property is retained only for backward compatibility.

$sel:hasTokenKey:APNSVoipChannelResponse', aPNSVoipChannelResponse_hasTokenKey - Specifies whether the APNs VoIP channel is configured to communicate with APNs by using APNs tokens. To provide an authentication key for APNs tokens, set the TokenKey property of the channel.

$sel:enabled:APNSVoipChannelResponse', aPNSVoipChannelResponse_enabled - Specifies whether the APNs VoIP channel is enabled for the application.

$sel:creationDate:APNSVoipChannelResponse', aPNSVoipChannelResponse_creationDate - The date and time when the APNs VoIP channel was enabled.

$sel:id:APNSVoipChannelResponse', aPNSVoipChannelResponse_id - (Deprecated) An identifier for the APNs VoIP channel. This property is retained only for backward compatibility.

$sel:version:APNSVoipChannelResponse', aPNSVoipChannelResponse_version - The current version of the APNs VoIP channel.

$sel:isArchived:APNSVoipChannelResponse', aPNSVoipChannelResponse_isArchived - Specifies whether the APNs VoIP channel is archived.

$sel:lastModifiedBy:APNSVoipChannelResponse', aPNSVoipChannelResponse_lastModifiedBy - The user who last modified the APNs VoIP channel.

$sel:platform:APNSVoipChannelResponse', aPNSVoipChannelResponse_platform - The type of messaging or notification platform for the channel. For the APNs VoIP channel, this value is APNS_VOIP.

aPNSVoipChannelResponse_applicationId :: Lens' APNSVoipChannelResponse (Maybe Text)

The unique identifier for the application that the APNs VoIP channel applies to.

aPNSVoipChannelResponse_lastModifiedDate :: Lens' APNSVoipChannelResponse (Maybe Text)

The date and time when the APNs VoIP channel was last modified.

aPNSVoipChannelResponse_defaultAuthenticationMethod :: Lens' APNSVoipChannelResponse (Maybe Text)

The default authentication method that Amazon Pinpoint uses to authenticate with APNs for this channel, key or certificate.

aPNSVoipChannelResponse_hasCredential :: Lens' APNSVoipChannelResponse (Maybe Bool)

(Not used) This property is retained only for backward compatibility.

aPNSVoipChannelResponse_hasTokenKey :: Lens' APNSVoipChannelResponse (Maybe Bool)

Specifies whether the APNs VoIP channel is configured to communicate with APNs by using APNs tokens. To provide an authentication key for APNs tokens, set the TokenKey property of the channel.

aPNSVoipChannelResponse_enabled :: Lens' APNSVoipChannelResponse (Maybe Bool)

Specifies whether the APNs VoIP channel is enabled for the application.

aPNSVoipChannelResponse_creationDate :: Lens' APNSVoipChannelResponse (Maybe Text)

The date and time when the APNs VoIP channel was enabled.

aPNSVoipChannelResponse_id :: Lens' APNSVoipChannelResponse (Maybe Text)

(Deprecated) An identifier for the APNs VoIP channel. This property is retained only for backward compatibility.

aPNSVoipChannelResponse_version :: Lens' APNSVoipChannelResponse (Maybe Int)

The current version of the APNs VoIP channel.

aPNSVoipChannelResponse_isArchived :: Lens' APNSVoipChannelResponse (Maybe Bool)

Specifies whether the APNs VoIP channel is archived.

aPNSVoipChannelResponse_lastModifiedBy :: Lens' APNSVoipChannelResponse (Maybe Text)

The user who last modified the APNs VoIP channel.

aPNSVoipChannelResponse_platform :: Lens' APNSVoipChannelResponse Text

The type of messaging or notification platform for the channel. For the APNs VoIP channel, this value is APNS_VOIP.