Network.AWS.Pinpoint.Types.APNSVoipSandboxChannelRequest

Description

Documentation

data APNSVoipSandboxChannelRequest

Specifies the status and settings of the APNs (Apple Push Notification service) VoIP sandbox channel for an application.

See: newAPNSVoipSandboxChannelRequest smart constructor.

Constructors

APNSVoipSandboxChannelRequest'

Fields

tokenKey :: Maybe Text

The authentication key to use for APNs tokens.

privateKey :: Maybe Text

The private key for the APNs client certificate that you want Amazon Pinpoint to use to communicate with the APNs sandbox environment.

enabled :: Maybe Bool

Specifies whether the APNs VoIP sandbox channel is enabled for the application.

teamId :: Maybe Text

The identifier that's assigned to your Apple developer account team. This identifier is used for APNs tokens.

bundleId :: Maybe Text

The bundle identifier that's assigned to your iOS app. This identifier is used for APNs tokens.

defaultAuthenticationMethod :: Maybe Text

The default authentication method that you want Amazon Pinpoint to use when authenticating with the APNs sandbox environment for this channel, key or certificate.

certificate :: Maybe Text

The APNs client certificate that you received from Apple, if you want Amazon Pinpoint to communicate with the APNs sandbox environment by using an APNs certificate.

tokenKeyId :: Maybe Text

The key identifier that's assigned to your APNs signing key, if you want Amazon Pinpoint to communicate with the APNs sandbox environment by using APNs tokens.

newAPNSVoipSandboxChannelRequest :: APNSVoipSandboxChannelRequest

Create a value of APNSVoipSandboxChannelRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tokenKey:APNSVoipSandboxChannelRequest', aPNSVoipSandboxChannelRequest_tokenKey - The authentication key to use for APNs tokens.

$sel:privateKey:APNSVoipSandboxChannelRequest', aPNSVoipSandboxChannelRequest_privateKey - The private key for the APNs client certificate that you want Amazon Pinpoint to use to communicate with the APNs sandbox environment.

$sel:enabled:APNSVoipSandboxChannelRequest', aPNSVoipSandboxChannelRequest_enabled - Specifies whether the APNs VoIP sandbox channel is enabled for the application.

$sel:teamId:APNSVoipSandboxChannelRequest', aPNSVoipSandboxChannelRequest_teamId - The identifier that's assigned to your Apple developer account team. This identifier is used for APNs tokens.

$sel:bundleId:APNSVoipSandboxChannelRequest', aPNSVoipSandboxChannelRequest_bundleId - The bundle identifier that's assigned to your iOS app. This identifier is used for APNs tokens.

$sel:defaultAuthenticationMethod:APNSVoipSandboxChannelRequest', aPNSVoipSandboxChannelRequest_defaultAuthenticationMethod - The default authentication method that you want Amazon Pinpoint to use when authenticating with the APNs sandbox environment for this channel, key or certificate.

$sel:certificate:APNSVoipSandboxChannelRequest', aPNSVoipSandboxChannelRequest_certificate - The APNs client certificate that you received from Apple, if you want Amazon Pinpoint to communicate with the APNs sandbox environment by using an APNs certificate.

$sel:tokenKeyId:APNSVoipSandboxChannelRequest', aPNSVoipSandboxChannelRequest_tokenKeyId - The key identifier that's assigned to your APNs signing key, if you want Amazon Pinpoint to communicate with the APNs sandbox environment by using APNs tokens.

aPNSVoipSandboxChannelRequest_tokenKey :: Lens' APNSVoipSandboxChannelRequest (Maybe Text)

The authentication key to use for APNs tokens.

aPNSVoipSandboxChannelRequest_privateKey :: Lens' APNSVoipSandboxChannelRequest (Maybe Text)

The private key for the APNs client certificate that you want Amazon Pinpoint to use to communicate with the APNs sandbox environment.

aPNSVoipSandboxChannelRequest_enabled :: Lens' APNSVoipSandboxChannelRequest (Maybe Bool)

Specifies whether the APNs VoIP sandbox channel is enabled for the application.

aPNSVoipSandboxChannelRequest_teamId :: Lens' APNSVoipSandboxChannelRequest (Maybe Text)

The identifier that's assigned to your Apple developer account team. This identifier is used for APNs tokens.

aPNSVoipSandboxChannelRequest_bundleId :: Lens' APNSVoipSandboxChannelRequest (Maybe Text)

The bundle identifier that's assigned to your iOS app. This identifier is used for APNs tokens.

aPNSVoipSandboxChannelRequest_defaultAuthenticationMethod :: Lens' APNSVoipSandboxChannelRequest (Maybe Text)

The default authentication method that you want Amazon Pinpoint to use when authenticating with the APNs sandbox environment for this channel, key or certificate.

aPNSVoipSandboxChannelRequest_certificate :: Lens' APNSVoipSandboxChannelRequest (Maybe Text)

The APNs client certificate that you received from Apple, if you want Amazon Pinpoint to communicate with the APNs sandbox environment by using an APNs certificate.

aPNSVoipSandboxChannelRequest_tokenKeyId :: Lens' APNSVoipSandboxChannelRequest (Maybe Text)

The key identifier that's assigned to your APNs signing key, if you want Amazon Pinpoint to communicate with the APNs sandbox environment by using APNs tokens.