Network.AWS.Pinpoint.Types.APNSVoipSandboxChannelResponse

Description

Documentation

data APNSVoipSandboxChannelResponse

Provides information about the status and settings of the APNs (Apple Push Notification service) VoIP sandbox channel for an application.

See: newAPNSVoipSandboxChannelResponse smart constructor.

Constructors

APNSVoipSandboxChannelResponse'

Fields

applicationId :: Maybe Text

The unique identifier for the application that the APNs VoIP sandbox channel applies to.

lastModifiedDate :: Maybe Text

The date and time when the APNs VoIP sandbox channel was last modified.

defaultAuthenticationMethod :: Maybe Text

The default authentication method that Amazon Pinpoint uses to authenticate with the APNs sandbox environment for this channel, key or certificate.

hasCredential :: Maybe Bool

(Not used) This property is retained only for backward compatibility.

hasTokenKey :: Maybe Bool

Specifies whether the APNs VoIP sandbox channel is configured to communicate with APNs by using APNs tokens. To provide an authentication key for APNs tokens, set the TokenKey property of the channel.

enabled :: Maybe Bool

Specifies whether the APNs VoIP sandbox channel is enabled for the application.

creationDate :: Maybe Text

The date and time when the APNs VoIP sandbox channel was enabled.

id :: Maybe Text

(Deprecated) An identifier for the APNs VoIP sandbox channel. This property is retained only for backward compatibility.

version :: Maybe Int

The current version of the APNs VoIP sandbox channel.

isArchived :: Maybe Bool

Specifies whether the APNs VoIP sandbox channel is archived.

lastModifiedBy :: Maybe Text

The user who last modified the APNs VoIP sandbox channel.

platform :: Text

The type of messaging or notification platform for the channel. For the APNs VoIP sandbox channel, this value is APNS_VOIP_SANDBOX.

newAPNSVoipSandboxChannelResponse

Arguments

:: Text
$sel:platform:APNSVoipSandboxChannelResponse'

-> APNSVoipSandboxChannelResponse

Create a value of APNSVoipSandboxChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:APNSVoipSandboxChannelResponse', aPNSVoipSandboxChannelResponse_applicationId - The unique identifier for the application that the APNs VoIP sandbox channel applies to.

$sel:lastModifiedDate:APNSVoipSandboxChannelResponse', aPNSVoipSandboxChannelResponse_lastModifiedDate - The date and time when the APNs VoIP sandbox channel was last modified.

$sel:defaultAuthenticationMethod:APNSVoipSandboxChannelResponse', aPNSVoipSandboxChannelResponse_defaultAuthenticationMethod - The default authentication method that Amazon Pinpoint uses to authenticate with the APNs sandbox environment for this channel, key or certificate.

$sel:hasCredential:APNSVoipSandboxChannelResponse', aPNSVoipSandboxChannelResponse_hasCredential - (Not used) This property is retained only for backward compatibility.

$sel:hasTokenKey:APNSVoipSandboxChannelResponse', aPNSVoipSandboxChannelResponse_hasTokenKey - Specifies whether the APNs VoIP sandbox channel is configured to communicate with APNs by using APNs tokens. To provide an authentication key for APNs tokens, set the TokenKey property of the channel.

$sel:enabled:APNSVoipSandboxChannelResponse', aPNSVoipSandboxChannelResponse_enabled - Specifies whether the APNs VoIP sandbox channel is enabled for the application.

$sel:creationDate:APNSVoipSandboxChannelResponse', aPNSVoipSandboxChannelResponse_creationDate - The date and time when the APNs VoIP sandbox channel was enabled.

$sel:id:APNSVoipSandboxChannelResponse', aPNSVoipSandboxChannelResponse_id - (Deprecated) An identifier for the APNs VoIP sandbox channel. This property is retained only for backward compatibility.

$sel:version:APNSVoipSandboxChannelResponse', aPNSVoipSandboxChannelResponse_version - The current version of the APNs VoIP sandbox channel.

$sel:isArchived:APNSVoipSandboxChannelResponse', aPNSVoipSandboxChannelResponse_isArchived - Specifies whether the APNs VoIP sandbox channel is archived.

$sel:lastModifiedBy:APNSVoipSandboxChannelResponse', aPNSVoipSandboxChannelResponse_lastModifiedBy - The user who last modified the APNs VoIP sandbox channel.

$sel:platform:APNSVoipSandboxChannelResponse', aPNSVoipSandboxChannelResponse_platform - The type of messaging or notification platform for the channel. For the APNs VoIP sandbox channel, this value is APNS_VOIP_SANDBOX.

aPNSVoipSandboxChannelResponse_applicationId :: Lens' APNSVoipSandboxChannelResponse (Maybe Text)

The unique identifier for the application that the APNs VoIP sandbox channel applies to.

aPNSVoipSandboxChannelResponse_lastModifiedDate :: Lens' APNSVoipSandboxChannelResponse (Maybe Text)

The date and time when the APNs VoIP sandbox channel was last modified.

aPNSVoipSandboxChannelResponse_defaultAuthenticationMethod :: Lens' APNSVoipSandboxChannelResponse (Maybe Text)

The default authentication method that Amazon Pinpoint uses to authenticate with the APNs sandbox environment for this channel, key or certificate.

aPNSVoipSandboxChannelResponse_hasCredential :: Lens' APNSVoipSandboxChannelResponse (Maybe Bool)

(Not used) This property is retained only for backward compatibility.

aPNSVoipSandboxChannelResponse_hasTokenKey :: Lens' APNSVoipSandboxChannelResponse (Maybe Bool)

Specifies whether the APNs VoIP sandbox channel is configured to communicate with APNs by using APNs tokens. To provide an authentication key for APNs tokens, set the TokenKey property of the channel.

aPNSVoipSandboxChannelResponse_enabled :: Lens' APNSVoipSandboxChannelResponse (Maybe Bool)

Specifies whether the APNs VoIP sandbox channel is enabled for the application.

aPNSVoipSandboxChannelResponse_creationDate :: Lens' APNSVoipSandboxChannelResponse (Maybe Text)

The date and time when the APNs VoIP sandbox channel was enabled.

aPNSVoipSandboxChannelResponse_id :: Lens' APNSVoipSandboxChannelResponse (Maybe Text)

(Deprecated) An identifier for the APNs VoIP sandbox channel. This property is retained only for backward compatibility.

aPNSVoipSandboxChannelResponse_version :: Lens' APNSVoipSandboxChannelResponse (Maybe Int)

The current version of the APNs VoIP sandbox channel.

aPNSVoipSandboxChannelResponse_isArchived :: Lens' APNSVoipSandboxChannelResponse (Maybe Bool)

Specifies whether the APNs VoIP sandbox channel is archived.

aPNSVoipSandboxChannelResponse_lastModifiedBy :: Lens' APNSVoipSandboxChannelResponse (Maybe Text)

The user who last modified the APNs VoIP sandbox channel.

aPNSVoipSandboxChannelResponse_platform :: Lens' APNSVoipSandboxChannelResponse Text

The type of messaging or notification platform for the channel. For the APNs VoIP sandbox channel, this value is APNS_VOIP_SANDBOX.