Network.AWS.Pinpoint.Types.Activity

Description

Documentation

data Activity

Specifies the configuration and other settings for an activity in a journey.

See: newActivity smart constructor.

Constructors

Activity'

Fields

conditionalSplit :: Maybe ConditionalSplitActivity

The settings for a yes/no split activity. This type of activity sends participants down one of two paths in a journey, based on conditions that you specify.

push :: Maybe PushMessageActivity

The settings for a push notification activity. This type of activity sends a push notification to participants.

wait :: Maybe WaitActivity

The settings for a wait activity. This type of activity waits for a certain amount of time or until a specific date and time before moving participants to the next activity in a journey.

custom :: Maybe CustomMessageActivity

The settings for a custom message activity. This type of activity calls an AWS Lambda function or web hook that sends messages to participants.

multiCondition :: Maybe MultiConditionalSplitActivity

The settings for a multivariate split activity. This type of activity sends participants down one of as many as five paths (including a default Else path) in a journey, based on conditions that you specify.

email :: Maybe EmailMessageActivity

The settings for an email activity. This type of activity sends an email message to participants.

holdout :: Maybe HoldoutActivity

The settings for a holdout activity. This type of activity stops a journey for a specified percentage of participants.

randomSplit :: Maybe RandomSplitActivity

The settings for a random split activity. This type of activity randomly sends specified percentages of participants down one of as many as five paths in a journey, based on conditions that you specify.

description :: Maybe Text

The custom description of the activity.

contactCenter :: Maybe ContactCenterActivity

The settings for a connect activity. This type of activity initiates a contact center call to participants.

sms :: Maybe SMSMessageActivity

The settings for an SMS activity. This type of activity sends a text message to participants.

newActivity :: Activity

Create a value of Activity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:conditionalSplit:Activity', activity_conditionalSplit - The settings for a yes/no split activity. This type of activity sends participants down one of two paths in a journey, based on conditions that you specify.

$sel:push:Activity', activity_push - The settings for a push notification activity. This type of activity sends a push notification to participants.

$sel:wait:Activity', activity_wait - The settings for a wait activity. This type of activity waits for a certain amount of time or until a specific date and time before moving participants to the next activity in a journey.

$sel:custom:Activity', activity_custom - The settings for a custom message activity. This type of activity calls an AWS Lambda function or web hook that sends messages to participants.

$sel:multiCondition:Activity', activity_multiCondition - The settings for a multivariate split activity. This type of activity sends participants down one of as many as five paths (including a default Else path) in a journey, based on conditions that you specify.

$sel:email:Activity', activity_email - The settings for an email activity. This type of activity sends an email message to participants.

$sel:holdout:Activity', activity_holdout - The settings for a holdout activity. This type of activity stops a journey for a specified percentage of participants.

$sel:randomSplit:Activity', activity_randomSplit - The settings for a random split activity. This type of activity randomly sends specified percentages of participants down one of as many as five paths in a journey, based on conditions that you specify.

$sel:description:Activity', activity_description - The custom description of the activity.

$sel:contactCenter:Activity', activity_contactCenter - The settings for a connect activity. This type of activity initiates a contact center call to participants.

$sel:sms:Activity', activity_sms - The settings for an SMS activity. This type of activity sends a text message to participants.

activity_conditionalSplit :: Lens' Activity (Maybe ConditionalSplitActivity)

The settings for a yes/no split activity. This type of activity sends participants down one of two paths in a journey, based on conditions that you specify.

activity_push :: Lens' Activity (Maybe PushMessageActivity)

The settings for a push notification activity. This type of activity sends a push notification to participants.

activity_wait :: Lens' Activity (Maybe WaitActivity)

The settings for a wait activity. This type of activity waits for a certain amount of time or until a specific date and time before moving participants to the next activity in a journey.

activity_custom :: Lens' Activity (Maybe CustomMessageActivity)

The settings for a custom message activity. This type of activity calls an AWS Lambda function or web hook that sends messages to participants.

activity_multiCondition :: Lens' Activity (Maybe MultiConditionalSplitActivity)

The settings for a multivariate split activity. This type of activity sends participants down one of as many as five paths (including a default Else path) in a journey, based on conditions that you specify.

activity_email :: Lens' Activity (Maybe EmailMessageActivity)

The settings for an email activity. This type of activity sends an email message to participants.

activity_holdout :: Lens' Activity (Maybe HoldoutActivity)

The settings for a holdout activity. This type of activity stops a journey for a specified percentage of participants.

activity_randomSplit :: Lens' Activity (Maybe RandomSplitActivity)

The settings for a random split activity. This type of activity randomly sends specified percentages of participants down one of as many as five paths in a journey, based on conditions that you specify.

activity_description :: Lens' Activity (Maybe Text)

The custom description of the activity.

activity_contactCenter :: Lens' Activity (Maybe ContactCenterActivity)

The settings for a connect activity. This type of activity initiates a contact center call to participants.

activity_sms :: Lens' Activity (Maybe SMSMessageActivity)

The settings for an SMS activity. This type of activity sends a text message to participants.