Network.AWS.Pinpoint.Types.AddressConfiguration

Description

Documentation

data AddressConfiguration

Specifies address-based configuration settings for a message that's sent directly to an endpoint.

See: newAddressConfiguration smart constructor.

Constructors

AddressConfiguration'

Fields

substitutions :: Maybe (HashMap Text [Text])

A map of the message variables to merge with the variables specified by properties of the DefaultMessage object. The variables specified in this map take precedence over all other variables.

titleOverride :: Maybe Text

The message title to use instead of the default message title. This value overrides the default message title.

context :: Maybe (HashMap Text Text)

An object that maps custom attributes to attributes for the address and is attached to the message. Attribute names are case sensitive.

For a push notification, this payload is added to the data.pinpoint object. For an email or text message, this payload is added to email/SMS delivery receipt event attributes.

rawContent :: Maybe Text

The raw, JSON-formatted string to use as the payload for the message. If specified, this value overrides all other values for the message.

bodyOverride :: Maybe Text

The message body to use instead of the default message body. This value overrides the default message body.

channelType :: Maybe ChannelType

The channel to use when sending the message.

newAddressConfiguration :: AddressConfiguration

Create a value of AddressConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:substitutions:AddressConfiguration', addressConfiguration_substitutions - A map of the message variables to merge with the variables specified by properties of the DefaultMessage object. The variables specified in this map take precedence over all other variables.

$sel:titleOverride:AddressConfiguration', addressConfiguration_titleOverride - The message title to use instead of the default message title. This value overrides the default message title.

$sel:context:AddressConfiguration', addressConfiguration_context - An object that maps custom attributes to attributes for the address and is attached to the message. Attribute names are case sensitive.

For a push notification, this payload is added to the data.pinpoint object. For an email or text message, this payload is added to email/SMS delivery receipt event attributes.

$sel:rawContent:AddressConfiguration', addressConfiguration_rawContent - The raw, JSON-formatted string to use as the payload for the message. If specified, this value overrides all other values for the message.

$sel:bodyOverride:AddressConfiguration', addressConfiguration_bodyOverride - The message body to use instead of the default message body. This value overrides the default message body.

$sel:channelType:AddressConfiguration', addressConfiguration_channelType - The channel to use when sending the message.

addressConfiguration_substitutions :: Lens' AddressConfiguration (Maybe (HashMap Text [Text]))

A map of the message variables to merge with the variables specified by properties of the DefaultMessage object. The variables specified in this map take precedence over all other variables.

addressConfiguration_titleOverride :: Lens' AddressConfiguration (Maybe Text)

The message title to use instead of the default message title. This value overrides the default message title.

addressConfiguration_context :: Lens' AddressConfiguration (Maybe (HashMap Text Text))

An object that maps custom attributes to attributes for the address and is attached to the message. Attribute names are case sensitive.

For a push notification, this payload is added to the data.pinpoint object. For an email or text message, this payload is added to email/SMS delivery receipt event attributes.

addressConfiguration_rawContent :: Lens' AddressConfiguration (Maybe Text)

The raw, JSON-formatted string to use as the payload for the message. If specified, this value overrides all other values for the message.

addressConfiguration_bodyOverride :: Lens' AddressConfiguration (Maybe Text)

The message body to use instead of the default message body. This value overrides the default message body.

addressConfiguration_channelType :: Lens' AddressConfiguration (Maybe ChannelType)

The channel to use when sending the message.