Network.AWS.Pinpoint.Types.AndroidPushNotificationTemplate

Description

Documentation

data AndroidPushNotificationTemplate

Specifies channel-specific content and settings for a message template that can be used in push notifications that are sent through the ADM (Amazon Device Messaging), Baidu (Baidu Cloud Push), or GCM (Firebase Cloud Messaging, formerly Google Cloud Messaging) channel.

See: newAndroidPushNotificationTemplate smart constructor.

Constructors

AndroidPushNotificationTemplate'

Fields

imageIconUrl :: Maybe Text

The URL of the large icon image to display in the content view of a push notification that's based on the message template.

rawContent :: Maybe Text

The raw, JSON-formatted string to use as the payload for a push notification that's based on the message template. If specified, this value overrides all other content for the message template.

smallImageIconUrl :: Maybe Text

The URL of the small icon image to display in the status bar and the content view of a push notification that's based on the message template.

body :: Maybe Text

The message body to use in a push notification that's based on the message template.

url :: Maybe Text

The URL to open in a recipient's default mobile browser, if a recipient taps a push notification that's based on the message template and the value of the Action property is URL.

sound :: Maybe Text

The sound to play when a recipient receives a push notification that's based on the message template. You can use the default stream or specify the file name of a sound resource that's bundled in your app. On an Android platform, the sound file must reside in /res/raw/.

action :: Maybe Action

The action to occur if a recipient taps a push notification that's based on the message template. Valid values are:

OPEN_APP - Your app opens or it becomes the foreground app if it was sent to the background. This is the default action.

DEEP_LINK - Your app opens and displays a designated user interface in the app. This action uses the deep-linking features of the Android platform.

URL - The default mobile browser on the recipient's device opens and loads the web page at a URL that you specify.

imageUrl :: Maybe Text

The URL of an image to display in a push notification that's based on the message template.

title :: Maybe Text

The title to use in a push notification that's based on the message template. This title appears above the notification message on a recipient's device.

newAndroidPushNotificationTemplate :: AndroidPushNotificationTemplate

Create a value of AndroidPushNotificationTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:imageIconUrl:AndroidPushNotificationTemplate', androidPushNotificationTemplate_imageIconUrl - The URL of the large icon image to display in the content view of a push notification that's based on the message template.

$sel:rawContent:AndroidPushNotificationTemplate', androidPushNotificationTemplate_rawContent - The raw, JSON-formatted string to use as the payload for a push notification that's based on the message template. If specified, this value overrides all other content for the message template.

$sel:smallImageIconUrl:AndroidPushNotificationTemplate', androidPushNotificationTemplate_smallImageIconUrl - The URL of the small icon image to display in the status bar and the content view of a push notification that's based on the message template.

$sel:body:AndroidPushNotificationTemplate', androidPushNotificationTemplate_body - The message body to use in a push notification that's based on the message template.

$sel:url:AndroidPushNotificationTemplate', androidPushNotificationTemplate_url - The URL to open in a recipient's default mobile browser, if a recipient taps a push notification that's based on the message template and the value of the Action property is URL.

$sel:sound:AndroidPushNotificationTemplate', androidPushNotificationTemplate_sound - The sound to play when a recipient receives a push notification that's based on the message template. You can use the default stream or specify the file name of a sound resource that's bundled in your app. On an Android platform, the sound file must reside in /res/raw/.

$sel:action:AndroidPushNotificationTemplate', androidPushNotificationTemplate_action - The action to occur if a recipient taps a push notification that's based on the message template. Valid values are:

OPEN_APP - Your app opens or it becomes the foreground app if it was sent to the background. This is the default action.

DEEP_LINK - Your app opens and displays a designated user interface in the app. This action uses the deep-linking features of the Android platform.

URL - The default mobile browser on the recipient's device opens and loads the web page at a URL that you specify.

$sel:imageUrl:AndroidPushNotificationTemplate', androidPushNotificationTemplate_imageUrl - The URL of an image to display in a push notification that's based on the message template.

$sel:title:AndroidPushNotificationTemplate', androidPushNotificationTemplate_title - The title to use in a push notification that's based on the message template. This title appears above the notification message on a recipient's device.

androidPushNotificationTemplate_imageIconUrl :: Lens' AndroidPushNotificationTemplate (Maybe Text)

The URL of the large icon image to display in the content view of a push notification that's based on the message template.

androidPushNotificationTemplate_rawContent :: Lens' AndroidPushNotificationTemplate (Maybe Text)

The raw, JSON-formatted string to use as the payload for a push notification that's based on the message template. If specified, this value overrides all other content for the message template.

androidPushNotificationTemplate_smallImageIconUrl :: Lens' AndroidPushNotificationTemplate (Maybe Text)

The URL of the small icon image to display in the status bar and the content view of a push notification that's based on the message template.

androidPushNotificationTemplate_body :: Lens' AndroidPushNotificationTemplate (Maybe Text)

The message body to use in a push notification that's based on the message template.

androidPushNotificationTemplate_url :: Lens' AndroidPushNotificationTemplate (Maybe Text)

The URL to open in a recipient's default mobile browser, if a recipient taps a push notification that's based on the message template and the value of the Action property is URL.

androidPushNotificationTemplate_sound :: Lens' AndroidPushNotificationTemplate (Maybe Text)

The sound to play when a recipient receives a push notification that's based on the message template. You can use the default stream or specify the file name of a sound resource that's bundled in your app. On an Android platform, the sound file must reside in /res/raw/.

androidPushNotificationTemplate_action :: Lens' AndroidPushNotificationTemplate (Maybe Action)

The action to occur if a recipient taps a push notification that's based on the message template. Valid values are:

OPEN_APP - Your app opens or it becomes the foreground app if it was sent to the background. This is the default action.

DEEP_LINK - Your app opens and displays a designated user interface in the app. This action uses the deep-linking features of the Android platform.

URL - The default mobile browser on the recipient's device opens and loads the web page at a URL that you specify.

androidPushNotificationTemplate_imageUrl :: Lens' AndroidPushNotificationTemplate (Maybe Text)

The URL of an image to display in a push notification that's based on the message template.

androidPushNotificationTemplate_title :: Lens' AndroidPushNotificationTemplate (Maybe Text)

The title to use in a push notification that's based on the message template. This title appears above the notification message on a recipient's device.