Network.AWS.Pinpoint.Types.ApplicationSettingsResource

Description

Documentation

data ApplicationSettingsResource

Provides information about an application, including the default settings for an application.

See: newApplicationSettingsResource smart constructor.

Constructors

ApplicationSettingsResource'

Fields

lastModifiedDate :: Maybe Text

The date and time, in ISO 8601 format, when the application's settings were last modified.

campaignHook :: Maybe CampaignHook

The settings for the AWS Lambda function to invoke by default as a code hook for campaigns in the application. You can use this hook to customize segments that are used by campaigns in the application.

quietTime :: Maybe QuietTime

The default quiet time for campaigns in the application. Quiet time is a specific time range when messages aren't sent to endpoints, if all the following conditions are met:

The EndpointDemographic.Timezone property of the endpoint is set to a valid value.

The current time in the endpoint's time zone is later than or equal to the time specified by the QuietTime.Start property for the application (or a campaign or journey that has custom quiet time settings).

The current time in the endpoint's time zone is earlier than or equal to the time specified by the QuietTime.End property for the application (or a campaign or journey that has custom quiet time settings).

If any of the preceding conditions isn't met, the endpoint will receive messages from a campaign or journey, even if quiet time is enabled.

limits :: Maybe CampaignLimits

The default sending limits for campaigns in the application.

applicationId :: Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

newApplicationSettingsResource

Arguments

:: Text
$sel:applicationId:ApplicationSettingsResource'

-> ApplicationSettingsResource

Create a value of ApplicationSettingsResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:ApplicationSettingsResource', applicationSettingsResource_lastModifiedDate - The date and time, in ISO 8601 format, when the application's settings were last modified.

$sel:campaignHook:ApplicationSettingsResource', applicationSettingsResource_campaignHook - The settings for the AWS Lambda function to invoke by default as a code hook for campaigns in the application. You can use this hook to customize segments that are used by campaigns in the application.

$sel:quietTime:ApplicationSettingsResource', applicationSettingsResource_quietTime - The default quiet time for campaigns in the application. Quiet time is a specific time range when messages aren't sent to endpoints, if all the following conditions are met:

The EndpointDemographic.Timezone property of the endpoint is set to a valid value.

The current time in the endpoint's time zone is later than or equal to the time specified by the QuietTime.Start property for the application (or a campaign or journey that has custom quiet time settings).

The current time in the endpoint's time zone is earlier than or equal to the time specified by the QuietTime.End property for the application (or a campaign or journey that has custom quiet time settings).

If any of the preceding conditions isn't met, the endpoint will receive messages from a campaign or journey, even if quiet time is enabled.

$sel:limits:ApplicationSettingsResource', applicationSettingsResource_limits - The default sending limits for campaigns in the application.

$sel:applicationId:ApplicationSettingsResource', applicationSettingsResource_applicationId - The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.

applicationSettingsResource_lastModifiedDate :: Lens' ApplicationSettingsResource (Maybe Text)

The date and time, in ISO 8601 format, when the application's settings were last modified.

applicationSettingsResource_campaignHook :: Lens' ApplicationSettingsResource (Maybe CampaignHook)

The settings for the AWS Lambda function to invoke by default as a code hook for campaigns in the application. You can use this hook to customize segments that are used by campaigns in the application.

applicationSettingsResource_quietTime :: Lens' ApplicationSettingsResource (Maybe QuietTime)

The default quiet time for campaigns in the application. Quiet time is a specific time range when messages aren't sent to endpoints, if all the following conditions are met:

The EndpointDemographic.Timezone property of the endpoint is set to a valid value.

The current time in the endpoint's time zone is later than or equal to the time specified by the QuietTime.Start property for the application (or a campaign or journey that has custom quiet time settings).

The current time in the endpoint's time zone is earlier than or equal to the time specified by the QuietTime.End property for the application (or a campaign or journey that has custom quiet time settings).

If any of the preceding conditions isn't met, the endpoint will receive messages from a campaign or journey, even if quiet time is enabled.

applicationSettingsResource_limits :: Lens' ApplicationSettingsResource (Maybe CampaignLimits)

The default sending limits for campaigns in the application.

applicationSettingsResource_applicationId :: Lens' ApplicationSettingsResource Text

The unique identifier for the application. This identifier is displayed as the Project ID on the Amazon Pinpoint console.