Network.AWS.Pinpoint.Types.BaiduMessage

Description

Documentation

data BaiduMessage

Specifies the settings for a one-time message that's sent directly to an endpoint through the Baidu (Baidu Cloud Push) channel.

See: newBaiduMessage smart constructor.

Constructors

BaiduMessage'

Fields

silentPush :: Maybe Bool

Specifies whether the notification is a silent push notification, which is a push notification that doesn't display on a recipient's device. Silent push notifications can be used for cases such as updating an app's configuration or supporting phone home functionality.

imageIconUrl :: Maybe Text

The URL of the large icon image to display in the content view of the push notification.

body :: Maybe Text

The body of the notification message.

iconReference :: Maybe Text

The icon image name of the asset saved in your app.

data' :: Maybe (HashMap Text Text)

The JSON data payload to use for the push notification, if the notification is a silent push notification. This payload is added to the data.pinpoint.jsonBody object of the notification.

title :: Maybe Text

The title to display above the notification message on the recipient's device.

timeToLive :: Maybe Int

The amount of time, in seconds, that the Baidu Cloud Push service should store the message if the recipient's device is offline. The default value and maximum supported time is 604,800 seconds (7 days).

imageUrl :: Maybe Text

The URL of an image to display in the push notification.

substitutions :: Maybe (HashMap Text [Text])

The default message variables to use in the notification message. You can override the default variables with individual address variables.

action :: Maybe Action

The action to occur if the recipient taps the push notification. Valid values are:

OPEN_APP - Your app opens or it becomes the foreground app if it was sent to the background. This is the default action.

DEEP_LINK - Your app opens and displays a designated user interface in the app. This action uses the deep-linking features of the Android platform.

URL - The default mobile browser on the recipient's device opens and loads the web page at a URL that you specify.

url :: Maybe Text

The URL to open in the recipient's default mobile browser, if a recipient taps the push notification and the value of the Action property is URL.

sound :: Maybe Text

The sound to play when the recipient receives the push notification. You can use the default stream or specify the file name of a sound resource that's bundled in your app. On an Android platform, the sound file must reside in /res/raw/.

smallImageIconUrl :: Maybe Text

The URL of the small icon image to display in the status bar and the content view of the push notification.

rawContent :: Maybe Text

The raw, JSON-formatted string to use as the payload for the notification message. If specified, this value overrides all other content for the message.

newBaiduMessage :: BaiduMessage

Create a value of BaiduMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:silentPush:BaiduMessage', baiduMessage_silentPush - Specifies whether the notification is a silent push notification, which is a push notification that doesn't display on a recipient's device. Silent push notifications can be used for cases such as updating an app's configuration or supporting phone home functionality.

$sel:imageIconUrl:BaiduMessage', baiduMessage_imageIconUrl - The URL of the large icon image to display in the content view of the push notification.

$sel:body:BaiduMessage', baiduMessage_body - The body of the notification message.

$sel:iconReference:BaiduMessage', baiduMessage_iconReference - The icon image name of the asset saved in your app.

$sel:data':BaiduMessage', baiduMessage_data - The JSON data payload to use for the push notification, if the notification is a silent push notification. This payload is added to the data.pinpoint.jsonBody object of the notification.

$sel:title:BaiduMessage', baiduMessage_title - The title to display above the notification message on the recipient's device.

$sel:timeToLive:BaiduMessage', baiduMessage_timeToLive - The amount of time, in seconds, that the Baidu Cloud Push service should store the message if the recipient's device is offline. The default value and maximum supported time is 604,800 seconds (7 days).

$sel:imageUrl:BaiduMessage', baiduMessage_imageUrl - The URL of an image to display in the push notification.

$sel:substitutions:BaiduMessage', baiduMessage_substitutions - The default message variables to use in the notification message. You can override the default variables with individual address variables.

$sel:action:BaiduMessage', baiduMessage_action - The action to occur if the recipient taps the push notification. Valid values are:

OPEN_APP - Your app opens or it becomes the foreground app if it was sent to the background. This is the default action.

DEEP_LINK - Your app opens and displays a designated user interface in the app. This action uses the deep-linking features of the Android platform.

URL - The default mobile browser on the recipient's device opens and loads the web page at a URL that you specify.

$sel:url:BaiduMessage', baiduMessage_url - The URL to open in the recipient's default mobile browser, if a recipient taps the push notification and the value of the Action property is URL.

$sel:sound:BaiduMessage', baiduMessage_sound - The sound to play when the recipient receives the push notification. You can use the default stream or specify the file name of a sound resource that's bundled in your app. On an Android platform, the sound file must reside in /res/raw/.

$sel:smallImageIconUrl:BaiduMessage', baiduMessage_smallImageIconUrl - The URL of the small icon image to display in the status bar and the content view of the push notification.

$sel:rawContent:BaiduMessage', baiduMessage_rawContent - The raw, JSON-formatted string to use as the payload for the notification message. If specified, this value overrides all other content for the message.

baiduMessage_silentPush :: Lens' BaiduMessage (Maybe Bool)

Specifies whether the notification is a silent push notification, which is a push notification that doesn't display on a recipient's device. Silent push notifications can be used for cases such as updating an app's configuration or supporting phone home functionality.

baiduMessage_imageIconUrl :: Lens' BaiduMessage (Maybe Text)

The URL of the large icon image to display in the content view of the push notification.

baiduMessage_body :: Lens' BaiduMessage (Maybe Text)

The body of the notification message.

baiduMessage_iconReference :: Lens' BaiduMessage (Maybe Text)

The icon image name of the asset saved in your app.

baiduMessage_data :: Lens' BaiduMessage (Maybe (HashMap Text Text))

The JSON data payload to use for the push notification, if the notification is a silent push notification. This payload is added to the data.pinpoint.jsonBody object of the notification.

baiduMessage_title :: Lens' BaiduMessage (Maybe Text)

The title to display above the notification message on the recipient's device.

baiduMessage_timeToLive :: Lens' BaiduMessage (Maybe Int)

The amount of time, in seconds, that the Baidu Cloud Push service should store the message if the recipient's device is offline. The default value and maximum supported time is 604,800 seconds (7 days).

baiduMessage_imageUrl :: Lens' BaiduMessage (Maybe Text)

The URL of an image to display in the push notification.

baiduMessage_substitutions :: Lens' BaiduMessage (Maybe (HashMap Text [Text]))

The default message variables to use in the notification message. You can override the default variables with individual address variables.

baiduMessage_action :: Lens' BaiduMessage (Maybe Action)

The action to occur if the recipient taps the push notification. Valid values are:

OPEN_APP - Your app opens or it becomes the foreground app if it was sent to the background. This is the default action.

DEEP_LINK - Your app opens and displays a designated user interface in the app. This action uses the deep-linking features of the Android platform.

URL - The default mobile browser on the recipient's device opens and loads the web page at a URL that you specify.

baiduMessage_url :: Lens' BaiduMessage (Maybe Text)

The URL to open in the recipient's default mobile browser, if a recipient taps the push notification and the value of the Action property is URL.

baiduMessage_sound :: Lens' BaiduMessage (Maybe Text)

The sound to play when the recipient receives the push notification. You can use the default stream or specify the file name of a sound resource that's bundled in your app. On an Android platform, the sound file must reside in /res/raw/.

baiduMessage_smallImageIconUrl :: Lens' BaiduMessage (Maybe Text)

The URL of the small icon image to display in the status bar and the content view of the push notification.

baiduMessage_rawContent :: Lens' BaiduMessage (Maybe Text)

The raw, JSON-formatted string to use as the payload for the notification message. If specified, this value overrides all other content for the message.