Network.AWS.Pinpoint.Types.CampaignResponse

Description

Documentation

data CampaignResponse

Provides information about the status, configuration, and other settings for a campaign.

See: newCampaignResponse smart constructor.

Constructors

CampaignResponse'

Fields

customDeliveryConfiguration :: Maybe CustomDeliveryConfiguration

The delivery configuration settings for sending the campaign through a custom channel.

state :: Maybe CampaignState

The current status of the campaign.

priority :: Maybe Int

Defines the priority of the campaign, used to decide the order of messages displayed to user if there are multiple messages scheduled to be displayed at the same moment.

schedule :: Maybe Schedule

The schedule settings for the campaign.

templateConfiguration :: Maybe TemplateConfiguration

The message template that’s used for the campaign.

hook :: Maybe CampaignHook

The settings for the AWS Lambda function to use as a code hook for the campaign. You can use this hook to customize the segment that's used by the campaign.

treatmentName :: Maybe Text

The custom name of the default treatment for the campaign, if the campaign has multiple treatments. A treatment is a variation of a campaign that's used for A/B testing.

limits :: Maybe CampaignLimits

The messaging limits for the campaign.

isPaused :: Maybe Bool

Specifies whether the campaign is paused. A paused campaign doesn't run unless you resume it by changing this value to false.

defaultState :: Maybe CampaignState

The current status of the campaign's default treatment. This value exists only for campaigns that have more than one treatment.

name :: Maybe Text

The name of the campaign.

version :: Maybe Int

The version number of the campaign.

holdoutPercent :: Maybe Int

The allocated percentage of users (segment members) who shouldn't receive messages from the campaign.

treatmentDescription :: Maybe Text

The custom description of the default treatment for the campaign.

messageConfiguration :: Maybe MessageConfiguration

The message configuration settings for the campaign.

description :: Maybe Text

The custom description of the campaign.

additionalTreatments :: Maybe [TreatmentResource]

An array of responses, one for each treatment that you defined for the campaign, in addition to the default treatment.

tags :: Maybe (HashMap Text Text)

A string-to-string map of key-value pairs that identifies the tags that are associated with the campaign. Each tag consists of a required tag key and an associated tag value.

lastModifiedDate :: Text

The date, in ISO 8601 format, when the campaign was last modified.

creationDate :: Text

The date, in ISO 8601 format, when the campaign was created.

segmentId :: Text

The unique identifier for the segment that's associated with the campaign.

segmentVersion :: Int

The version number of the segment that's associated with the campaign.

id :: Text

The unique identifier for the campaign.

arn :: Text

The Amazon Resource Name (ARN) of the campaign.

applicationId :: Text

The unique identifier for the application that the campaign applies to.

newCampaignResponse

Arguments

:: Text
$sel:lastModifiedDate:CampaignResponse'

-> Text
$sel:creationDate:CampaignResponse'

-> Text
$sel:segmentId:CampaignResponse'

-> Int
$sel:segmentVersion:CampaignResponse'

-> Text
$sel:id:CampaignResponse'

-> Text
$sel:arn:CampaignResponse'

-> Text
$sel:applicationId:CampaignResponse'

-> CampaignResponse

Create a value of CampaignResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customDeliveryConfiguration:CampaignResponse', campaignResponse_customDeliveryConfiguration - The delivery configuration settings for sending the campaign through a custom channel.

$sel:state:CampaignResponse', campaignResponse_state - The current status of the campaign.

$sel:priority:CampaignResponse', campaignResponse_priority - Defines the priority of the campaign, used to decide the order of messages displayed to user if there are multiple messages scheduled to be displayed at the same moment.

$sel:schedule:CampaignResponse', campaignResponse_schedule - The schedule settings for the campaign.

$sel:templateConfiguration:CampaignResponse', campaignResponse_templateConfiguration - The message template that’s used for the campaign.

$sel:hook:CampaignResponse', campaignResponse_hook - The settings for the AWS Lambda function to use as a code hook for the campaign. You can use this hook to customize the segment that's used by the campaign.

$sel:treatmentName:CampaignResponse', campaignResponse_treatmentName - The custom name of the default treatment for the campaign, if the campaign has multiple treatments. A treatment is a variation of a campaign that's used for A/B testing.

$sel:limits:CampaignResponse', campaignResponse_limits - The messaging limits for the campaign.

$sel:isPaused:CampaignResponse', campaignResponse_isPaused - Specifies whether the campaign is paused. A paused campaign doesn't run unless you resume it by changing this value to false.

$sel:defaultState:CampaignResponse', campaignResponse_defaultState - The current status of the campaign's default treatment. This value exists only for campaigns that have more than one treatment.

$sel:name:CampaignResponse', campaignResponse_name - The name of the campaign.

$sel:version:CampaignResponse', campaignResponse_version - The version number of the campaign.

$sel:holdoutPercent:CampaignResponse', campaignResponse_holdoutPercent - The allocated percentage of users (segment members) who shouldn't receive messages from the campaign.

$sel:treatmentDescription:CampaignResponse', campaignResponse_treatmentDescription - The custom description of the default treatment for the campaign.

$sel:messageConfiguration:CampaignResponse', campaignResponse_messageConfiguration - The message configuration settings for the campaign.

$sel:description:CampaignResponse', campaignResponse_description - The custom description of the campaign.

$sel:additionalTreatments:CampaignResponse', campaignResponse_additionalTreatments - An array of responses, one for each treatment that you defined for the campaign, in addition to the default treatment.

$sel:tags:CampaignResponse', campaignResponse_tags - A string-to-string map of key-value pairs that identifies the tags that are associated with the campaign. Each tag consists of a required tag key and an associated tag value.

$sel:lastModifiedDate:CampaignResponse', campaignResponse_lastModifiedDate - The date, in ISO 8601 format, when the campaign was last modified.

$sel:creationDate:CampaignResponse', campaignResponse_creationDate - The date, in ISO 8601 format, when the campaign was created.

$sel:segmentId:CampaignResponse', campaignResponse_segmentId - The unique identifier for the segment that's associated with the campaign.

$sel:segmentVersion:CampaignResponse', campaignResponse_segmentVersion - The version number of the segment that's associated with the campaign.

$sel:id:CampaignResponse', campaignResponse_id - The unique identifier for the campaign.

$sel:arn:CampaignResponse', campaignResponse_arn - The Amazon Resource Name (ARN) of the campaign.

$sel:applicationId:CampaignResponse', campaignResponse_applicationId - The unique identifier for the application that the campaign applies to.

campaignResponse_customDeliveryConfiguration :: Lens' CampaignResponse (Maybe CustomDeliveryConfiguration)

The delivery configuration settings for sending the campaign through a custom channel.

campaignResponse_state :: Lens' CampaignResponse (Maybe CampaignState)

The current status of the campaign.

campaignResponse_priority :: Lens' CampaignResponse (Maybe Int)

Defines the priority of the campaign, used to decide the order of messages displayed to user if there are multiple messages scheduled to be displayed at the same moment.

campaignResponse_schedule :: Lens' CampaignResponse (Maybe Schedule)

The schedule settings for the campaign.

campaignResponse_templateConfiguration :: Lens' CampaignResponse (Maybe TemplateConfiguration)

The message template that’s used for the campaign.

campaignResponse_hook :: Lens' CampaignResponse (Maybe CampaignHook)

The settings for the AWS Lambda function to use as a code hook for the campaign. You can use this hook to customize the segment that's used by the campaign.

campaignResponse_treatmentName :: Lens' CampaignResponse (Maybe Text)

The custom name of the default treatment for the campaign, if the campaign has multiple treatments. A treatment is a variation of a campaign that's used for A/B testing.

campaignResponse_limits :: Lens' CampaignResponse (Maybe CampaignLimits)

The messaging limits for the campaign.

campaignResponse_isPaused :: Lens' CampaignResponse (Maybe Bool)

Specifies whether the campaign is paused. A paused campaign doesn't run unless you resume it by changing this value to false.

campaignResponse_defaultState :: Lens' CampaignResponse (Maybe CampaignState)

The current status of the campaign's default treatment. This value exists only for campaigns that have more than one treatment.

campaignResponse_name :: Lens' CampaignResponse (Maybe Text)

The name of the campaign.

campaignResponse_version :: Lens' CampaignResponse (Maybe Int)

The version number of the campaign.

campaignResponse_holdoutPercent :: Lens' CampaignResponse (Maybe Int)

The allocated percentage of users (segment members) who shouldn't receive messages from the campaign.

campaignResponse_treatmentDescription :: Lens' CampaignResponse (Maybe Text)

The custom description of the default treatment for the campaign.

campaignResponse_messageConfiguration :: Lens' CampaignResponse (Maybe MessageConfiguration)

The message configuration settings for the campaign.

campaignResponse_description :: Lens' CampaignResponse (Maybe Text)

The custom description of the campaign.

campaignResponse_additionalTreatments :: Lens' CampaignResponse (Maybe [TreatmentResource])

An array of responses, one for each treatment that you defined for the campaign, in addition to the default treatment.

campaignResponse_tags :: Lens' CampaignResponse (Maybe (HashMap Text Text))

A string-to-string map of key-value pairs that identifies the tags that are associated with the campaign. Each tag consists of a required tag key and an associated tag value.

campaignResponse_lastModifiedDate :: Lens' CampaignResponse Text

The date, in ISO 8601 format, when the campaign was last modified.

campaignResponse_creationDate :: Lens' CampaignResponse Text

The date, in ISO 8601 format, when the campaign was created.

campaignResponse_segmentId :: Lens' CampaignResponse Text

The unique identifier for the segment that's associated with the campaign.

campaignResponse_segmentVersion :: Lens' CampaignResponse Int

The version number of the segment that's associated with the campaign.

campaignResponse_id :: Lens' CampaignResponse Text

The unique identifier for the campaign.

campaignResponse_arn :: Lens' CampaignResponse Text

The Amazon Resource Name (ARN) of the campaign.

campaignResponse_applicationId :: Lens' CampaignResponse Text

The unique identifier for the application that the campaign applies to.