Network.AWS.Pinpoint.Types.Condition

Description

Documentation

data Condition

Specifies the conditions to evaluate for an activity in a journey, and how to evaluate those conditions.

See: newCondition smart constructor.

Constructors

Condition'

Fields

operator :: Maybe Operator

Specifies how to handle multiple conditions for the activity. For example, if you specify two conditions for an activity, whether both or only one of the conditions must be met for the activity to be performed.

conditions :: Maybe [SimpleCondition]

The conditions to evaluate for the activity.

newCondition :: Condition

Create a value of Condition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:operator:Condition', condition_operator - Specifies how to handle multiple conditions for the activity. For example, if you specify two conditions for an activity, whether both or only one of the conditions must be met for the activity to be performed.

$sel:conditions:Condition', condition_conditions - The conditions to evaluate for the activity.

condition_operator :: Lens' Condition (Maybe Operator)

Specifies how to handle multiple conditions for the activity. For example, if you specify two conditions for an activity, whether both or only one of the conditions must be met for the activity to be performed.

condition_conditions :: Lens' Condition (Maybe [SimpleCondition])

The conditions to evaluate for the activity.