Network.AWS.Pinpoint.Types.ConditionalSplitActivity

Description

Documentation

data ConditionalSplitActivity

Specifies the settings for a yes/no split activity in a journey. This type of activity sends participants down one of two paths in a journey, based on conditions that you specify.

To create yes/no split activities that send participants down different paths based on push notification events (such as Open or Received events), your mobile app has to specify the User ID and Endpoint ID values. For more information, see Integrating Amazon Pinpoint with your application in the Amazon Pinpoint Developer Guide.

See: newConditionalSplitActivity smart constructor.

Constructors

ConditionalSplitActivity'

Fields

trueActivity :: Maybe Text

The unique identifier for the activity to perform if the conditions are met.

condition :: Maybe Condition

The conditions that define the paths for the activity, and the relationship between the conditions.

evaluationWaitTime :: Maybe WaitTime

The amount of time to wait before determining whether the conditions are met, or the date and time when Amazon Pinpoint determines whether the conditions are met.

falseActivity :: Maybe Text

The unique identifier for the activity to perform if the conditions aren't met.

newConditionalSplitActivity :: ConditionalSplitActivity

Create a value of ConditionalSplitActivity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trueActivity:ConditionalSplitActivity', conditionalSplitActivity_trueActivity - The unique identifier for the activity to perform if the conditions are met.

$sel:condition:ConditionalSplitActivity', conditionalSplitActivity_condition - The conditions that define the paths for the activity, and the relationship between the conditions.

$sel:evaluationWaitTime:ConditionalSplitActivity', conditionalSplitActivity_evaluationWaitTime - The amount of time to wait before determining whether the conditions are met, or the date and time when Amazon Pinpoint determines whether the conditions are met.

$sel:falseActivity:ConditionalSplitActivity', conditionalSplitActivity_falseActivity - The unique identifier for the activity to perform if the conditions aren't met.

conditionalSplitActivity_trueActivity :: Lens' ConditionalSplitActivity (Maybe Text)

The unique identifier for the activity to perform if the conditions are met.

conditionalSplitActivity_condition :: Lens' ConditionalSplitActivity (Maybe Condition)

The conditions that define the paths for the activity, and the relationship between the conditions.

conditionalSplitActivity_evaluationWaitTime :: Lens' ConditionalSplitActivity (Maybe WaitTime)

The amount of time to wait before determining whether the conditions are met, or the date and time when Amazon Pinpoint determines whether the conditions are met.

conditionalSplitActivity_falseActivity :: Lens' ConditionalSplitActivity (Maybe Text)

The unique identifier for the activity to perform if the conditions aren't met.