Network.AWS.Pinpoint.Types.CreateApplicationRequest

Description

Documentation

data CreateApplicationRequest

Specifies the display name of an application and the tags to associate with the application.

See: newCreateApplicationRequest smart constructor.

Constructors

CreateApplicationRequest'

Fields

tags :: Maybe (HashMap Text Text)

A string-to-string map of key-value pairs that defines the tags to associate with the application. Each tag consists of a required tag key and an associated tag value.

name :: Text

The display name of the application. This name is displayed as the Project name on the Amazon Pinpoint console.

newCreateApplicationRequest

Arguments

:: Text
$sel:name:CreateApplicationRequest'

-> CreateApplicationRequest

Create a value of CreateApplicationRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateApplicationRequest', createApplicationRequest_tags - A string-to-string map of key-value pairs that defines the tags to associate with the application. Each tag consists of a required tag key and an associated tag value.

$sel:name:CreateApplicationRequest', createApplicationRequest_name - The display name of the application. This name is displayed as the Project name on the Amazon Pinpoint console.

createApplicationRequest_tags :: Lens' CreateApplicationRequest (Maybe (HashMap Text Text))

A string-to-string map of key-value pairs that defines the tags to associate with the application. Each tag consists of a required tag key and an associated tag value.

createApplicationRequest_name :: Lens' CreateApplicationRequest Text

The display name of the application. This name is displayed as the Project name on the Amazon Pinpoint console.