Network.AWS.Pinpoint.Types.DefaultPushNotificationMessage

Description

Documentation

data DefaultPushNotificationMessage

Specifies the default settings and content for a push notification that's sent directly to an endpoint.

See: newDefaultPushNotificationMessage smart constructor.

Constructors

DefaultPushNotificationMessage'

Fields

silentPush :: Maybe Bool

Specifies whether the default notification is a silent push notification, which is a push notification that doesn't display on a recipient's device. Silent push notifications can be used for cases such as updating an app's configuration or delivering messages to an in-app notification center.

body :: Maybe Text

The default body of the notification message.

data' :: Maybe (HashMap Text Text)

The JSON data payload to use for the default push notification, if the notification is a silent push notification. This payload is added to the data.pinpoint.jsonBody object of the notification.

title :: Maybe Text

The default title to display above the notification message on a recipient's device.

substitutions :: Maybe (HashMap Text [Text])

The default message variables to use in the notification message. You can override the default variables with individual address variables.

action :: Maybe Action

The default action to occur if a recipient taps the push notification. Valid values are:

OPEN_APP - Your app opens or it becomes the foreground app if it was sent to the background. This is the default action.

DEEP_LINK - Your app opens and displays a designated user interface in the app. This setting uses the deep-linking features of the iOS and Android platforms.

URL - The default mobile browser on the recipient's device opens and loads the web page at a URL that you specify.

url :: Maybe Text

The default URL to open in a recipient's default mobile browser, if a recipient taps the push notification and the value of the Action property is URL.

newDefaultPushNotificationMessage :: DefaultPushNotificationMessage

Create a value of DefaultPushNotificationMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:silentPush:DefaultPushNotificationMessage', defaultPushNotificationMessage_silentPush - Specifies whether the default notification is a silent push notification, which is a push notification that doesn't display on a recipient's device. Silent push notifications can be used for cases such as updating an app's configuration or delivering messages to an in-app notification center.

$sel:body:DefaultPushNotificationMessage', defaultPushNotificationMessage_body - The default body of the notification message.

$sel:data':DefaultPushNotificationMessage', defaultPushNotificationMessage_data - The JSON data payload to use for the default push notification, if the notification is a silent push notification. This payload is added to the data.pinpoint.jsonBody object of the notification.

$sel:title:DefaultPushNotificationMessage', defaultPushNotificationMessage_title - The default title to display above the notification message on a recipient's device.

$sel:substitutions:DefaultPushNotificationMessage', defaultPushNotificationMessage_substitutions - The default message variables to use in the notification message. You can override the default variables with individual address variables.

$sel:action:DefaultPushNotificationMessage', defaultPushNotificationMessage_action - The default action to occur if a recipient taps the push notification. Valid values are:

OPEN_APP - Your app opens or it becomes the foreground app if it was sent to the background. This is the default action.

DEEP_LINK - Your app opens and displays a designated user interface in the app. This setting uses the deep-linking features of the iOS and Android platforms.

URL - The default mobile browser on the recipient's device opens and loads the web page at a URL that you specify.

$sel:url:DefaultPushNotificationMessage', defaultPushNotificationMessage_url - The default URL to open in a recipient's default mobile browser, if a recipient taps the push notification and the value of the Action property is URL.

defaultPushNotificationMessage_silentPush :: Lens' DefaultPushNotificationMessage (Maybe Bool)

Specifies whether the default notification is a silent push notification, which is a push notification that doesn't display on a recipient's device. Silent push notifications can be used for cases such as updating an app's configuration or delivering messages to an in-app notification center.

defaultPushNotificationMessage_body :: Lens' DefaultPushNotificationMessage (Maybe Text)

The default body of the notification message.

defaultPushNotificationMessage_data :: Lens' DefaultPushNotificationMessage (Maybe (HashMap Text Text))

The JSON data payload to use for the default push notification, if the notification is a silent push notification. This payload is added to the data.pinpoint.jsonBody object of the notification.

defaultPushNotificationMessage_title :: Lens' DefaultPushNotificationMessage (Maybe Text)

The default title to display above the notification message on a recipient's device.

defaultPushNotificationMessage_substitutions :: Lens' DefaultPushNotificationMessage (Maybe (HashMap Text [Text]))

The default message variables to use in the notification message. You can override the default variables with individual address variables.

defaultPushNotificationMessage_action :: Lens' DefaultPushNotificationMessage (Maybe Action)

The default action to occur if a recipient taps the push notification. Valid values are:

OPEN_APP - Your app opens or it becomes the foreground app if it was sent to the background. This is the default action.

DEEP_LINK - Your app opens and displays a designated user interface in the app. This setting uses the deep-linking features of the iOS and Android platforms.

URL - The default mobile browser on the recipient's device opens and loads the web page at a URL that you specify.

defaultPushNotificationMessage_url :: Lens' DefaultPushNotificationMessage (Maybe Text)

The default URL to open in a recipient's default mobile browser, if a recipient taps the push notification and the value of the Action property is URL.