Network.AWS.Pinpoint.Types.DefaultPushNotificationTemplate

Description

Documentation

data DefaultPushNotificationTemplate

Specifies the default settings and content for a message template that can be used in messages that are sent through a push notification channel.

See: newDefaultPushNotificationTemplate smart constructor.

Constructors

DefaultPushNotificationTemplate'

Fields

body :: Maybe Text

The message body to use in push notifications that are based on the message template.

title :: Maybe Text

The title to use in push notifications that are based on the message template. This title appears above the notification message on a recipient's device.

action :: Maybe Action

The action to occur if a recipient taps a push notification that's based on the message template. Valid values are:

OPEN_APP - Your app opens or it becomes the foreground app if it was sent to the background. This is the default action.

DEEP_LINK - Your app opens and displays a designated user interface in the app. This setting uses the deep-linking features of the iOS and Android platforms.

URL - The default mobile browser on the recipient's device opens and loads the web page at a URL that you specify.

url :: Maybe Text

The URL to open in a recipient's default mobile browser, if a recipient taps a push notification that's based on the message template and the value of the Action property is URL.

sound :: Maybe Text

The sound to play when a recipient receives a push notification that's based on the message template. You can use the default stream or specify the file name of a sound resource that's bundled in your app. On an Android platform, the sound file must reside in /res/raw/.

For an iOS platform, this value is the key for the name of a sound file in your app's main bundle or the Library/Sounds folder in your app's data container. If the sound file can't be found or you specify default for the value, the system plays the default alert sound.

newDefaultPushNotificationTemplate :: DefaultPushNotificationTemplate

Create a value of DefaultPushNotificationTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:body:DefaultPushNotificationTemplate', defaultPushNotificationTemplate_body - The message body to use in push notifications that are based on the message template.

$sel:title:DefaultPushNotificationTemplate', defaultPushNotificationTemplate_title - The title to use in push notifications that are based on the message template. This title appears above the notification message on a recipient's device.

$sel:action:DefaultPushNotificationTemplate', defaultPushNotificationTemplate_action - The action to occur if a recipient taps a push notification that's based on the message template. Valid values are:

OPEN_APP - Your app opens or it becomes the foreground app if it was sent to the background. This is the default action.

DEEP_LINK - Your app opens and displays a designated user interface in the app. This setting uses the deep-linking features of the iOS and Android platforms.

URL - The default mobile browser on the recipient's device opens and loads the web page at a URL that you specify.

$sel:url:DefaultPushNotificationTemplate', defaultPushNotificationTemplate_url - The URL to open in a recipient's default mobile browser, if a recipient taps a push notification that's based on the message template and the value of the Action property is URL.

$sel:sound:DefaultPushNotificationTemplate', defaultPushNotificationTemplate_sound - The sound to play when a recipient receives a push notification that's based on the message template. You can use the default stream or specify the file name of a sound resource that's bundled in your app. On an Android platform, the sound file must reside in /res/raw/.

For an iOS platform, this value is the key for the name of a sound file in your app's main bundle or the Library/Sounds folder in your app's data container. If the sound file can't be found or you specify default for the value, the system plays the default alert sound.

defaultPushNotificationTemplate_body :: Lens' DefaultPushNotificationTemplate (Maybe Text)

The message body to use in push notifications that are based on the message template.

defaultPushNotificationTemplate_title :: Lens' DefaultPushNotificationTemplate (Maybe Text)

The title to use in push notifications that are based on the message template. This title appears above the notification message on a recipient's device.

defaultPushNotificationTemplate_action :: Lens' DefaultPushNotificationTemplate (Maybe Action)

The action to occur if a recipient taps a push notification that's based on the message template. Valid values are:

OPEN_APP - Your app opens or it becomes the foreground app if it was sent to the background. This is the default action.

DEEP_LINK - Your app opens and displays a designated user interface in the app. This setting uses the deep-linking features of the iOS and Android platforms.

URL - The default mobile browser on the recipient's device opens and loads the web page at a URL that you specify.

defaultPushNotificationTemplate_url :: Lens' DefaultPushNotificationTemplate (Maybe Text)

The URL to open in a recipient's default mobile browser, if a recipient taps a push notification that's based on the message template and the value of the Action property is URL.

defaultPushNotificationTemplate_sound :: Lens' DefaultPushNotificationTemplate (Maybe Text)

The sound to play when a recipient receives a push notification that's based on the message template. You can use the default stream or specify the file name of a sound resource that's bundled in your app. On an Android platform, the sound file must reside in /res/raw/.

For an iOS platform, this value is the key for the name of a sound file in your app's main bundle or the Library/Sounds folder in your app's data container. If the sound file can't be found or you specify default for the value, the system plays the default alert sound.