Network.AWS.Pinpoint.Types.DirectMessageConfiguration

Description

Documentation

data DirectMessageConfiguration

Specifies the settings and content for the default message and any default messages that you tailored for specific channels.

See: newDirectMessageConfiguration smart constructor.

Constructors

DirectMessageConfiguration'

Fields

aDMMessage :: Maybe ADMMessage

The default push notification message for the ADM (Amazon Device Messaging) channel. This message overrides the default push notification message (DefaultPushNotificationMessage).

defaultMessage :: Maybe DefaultMessage

The default message for all channels.

voiceMessage :: Maybe VoiceMessage

The default message for the voice channel. This message overrides the default message (DefaultMessage).

gCMMessage :: Maybe GCMMessage

The default push notification message for the GCM channel, which is used to send notifications through the Firebase Cloud Messaging (FCM), formerly Google Cloud Messaging (GCM), service. This message overrides the default push notification message (DefaultPushNotificationMessage).

emailMessage :: Maybe EmailMessage

The default message for the email channel. This message overrides the default message (DefaultMessage).

aPNSMessage :: Maybe APNSMessage

The default push notification message for the APNs (Apple Push Notification service) channel. This message overrides the default push notification message (DefaultPushNotificationMessage).

sMSMessage :: Maybe SMSMessage

The default message for the SMS channel. This message overrides the default message (DefaultMessage).

defaultPushNotificationMessage :: Maybe DefaultPushNotificationMessage

The default push notification message for all push notification channels.

baiduMessage :: Maybe BaiduMessage

The default push notification message for the Baidu (Baidu Cloud Push) channel. This message overrides the default push notification message (DefaultPushNotificationMessage).

newDirectMessageConfiguration :: DirectMessageConfiguration

Create a value of DirectMessageConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aDMMessage:DirectMessageConfiguration', directMessageConfiguration_aDMMessage - The default push notification message for the ADM (Amazon Device Messaging) channel. This message overrides the default push notification message (DefaultPushNotificationMessage).

$sel:defaultMessage:DirectMessageConfiguration', directMessageConfiguration_defaultMessage - The default message for all channels.

$sel:voiceMessage:DirectMessageConfiguration', directMessageConfiguration_voiceMessage - The default message for the voice channel. This message overrides the default message (DefaultMessage).

$sel:gCMMessage:DirectMessageConfiguration', directMessageConfiguration_gCMMessage - The default push notification message for the GCM channel, which is used to send notifications through the Firebase Cloud Messaging (FCM), formerly Google Cloud Messaging (GCM), service. This message overrides the default push notification message (DefaultPushNotificationMessage).

$sel:emailMessage:DirectMessageConfiguration', directMessageConfiguration_emailMessage - The default message for the email channel. This message overrides the default message (DefaultMessage).

$sel:aPNSMessage:DirectMessageConfiguration', directMessageConfiguration_aPNSMessage - The default push notification message for the APNs (Apple Push Notification service) channel. This message overrides the default push notification message (DefaultPushNotificationMessage).

$sel:sMSMessage:DirectMessageConfiguration', directMessageConfiguration_sMSMessage - The default message for the SMS channel. This message overrides the default message (DefaultMessage).

$sel:defaultPushNotificationMessage:DirectMessageConfiguration', directMessageConfiguration_defaultPushNotificationMessage - The default push notification message for all push notification channels.

$sel:baiduMessage:DirectMessageConfiguration', directMessageConfiguration_baiduMessage - The default push notification message for the Baidu (Baidu Cloud Push) channel. This message overrides the default push notification message (DefaultPushNotificationMessage).

directMessageConfiguration_aDMMessage :: Lens' DirectMessageConfiguration (Maybe ADMMessage)

The default push notification message for the ADM (Amazon Device Messaging) channel. This message overrides the default push notification message (DefaultPushNotificationMessage).

directMessageConfiguration_defaultMessage :: Lens' DirectMessageConfiguration (Maybe DefaultMessage)

The default message for all channels.

directMessageConfiguration_voiceMessage :: Lens' DirectMessageConfiguration (Maybe VoiceMessage)

The default message for the voice channel. This message overrides the default message (DefaultMessage).

directMessageConfiguration_gCMMessage :: Lens' DirectMessageConfiguration (Maybe GCMMessage)

The default push notification message for the GCM channel, which is used to send notifications through the Firebase Cloud Messaging (FCM), formerly Google Cloud Messaging (GCM), service. This message overrides the default push notification message (DefaultPushNotificationMessage).

directMessageConfiguration_emailMessage :: Lens' DirectMessageConfiguration (Maybe EmailMessage)

The default message for the email channel. This message overrides the default message (DefaultMessage).

directMessageConfiguration_aPNSMessage :: Lens' DirectMessageConfiguration (Maybe APNSMessage)

The default push notification message for the APNs (Apple Push Notification service) channel. This message overrides the default push notification message (DefaultPushNotificationMessage).

directMessageConfiguration_sMSMessage :: Lens' DirectMessageConfiguration (Maybe SMSMessage)

The default message for the SMS channel. This message overrides the default message (DefaultMessage).

directMessageConfiguration_defaultPushNotificationMessage :: Lens' DirectMessageConfiguration (Maybe DefaultPushNotificationMessage)

The default push notification message for all push notification channels.

directMessageConfiguration_baiduMessage :: Lens' DirectMessageConfiguration (Maybe BaiduMessage)

The default push notification message for the Baidu (Baidu Cloud Push) channel. This message overrides the default push notification message (DefaultPushNotificationMessage).