Network.AWS.Pinpoint.Types.EmailChannelRequest

Description

Documentation

data EmailChannelRequest

Specifies the status and settings of the email channel for an application.

See: newEmailChannelRequest smart constructor.

Constructors

EmailChannelRequest'

Fields

enabled :: Maybe Bool

Specifies whether to enable the email channel for the application.

configurationSet :: Maybe Text

The Amazon SES configuration set that you want to apply to messages that you send through the channel.

roleArn :: Maybe Text

The ARN of the AWS Identity and Access Management (IAM) role that you want Amazon Pinpoint to use when it submits email-related event data for the channel.

fromAddress :: Text

The verified email address that you want to send email from when you send email through the channel.

identity :: Text

The Amazon Resource Name (ARN) of the identity, verified with Amazon Simple Email Service (Amazon SES), that you want to use when you send email through the channel.

newEmailChannelRequest

Arguments

:: Text
$sel:fromAddress:EmailChannelRequest'

-> Text
$sel:identity:EmailChannelRequest'

-> EmailChannelRequest

Create a value of EmailChannelRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:EmailChannelRequest', emailChannelRequest_enabled - Specifies whether to enable the email channel for the application.

$sel:configurationSet:EmailChannelRequest', emailChannelRequest_configurationSet - The Amazon SES configuration set that you want to apply to messages that you send through the channel.

$sel:roleArn:EmailChannelRequest', emailChannelRequest_roleArn - The ARN of the AWS Identity and Access Management (IAM) role that you want Amazon Pinpoint to use when it submits email-related event data for the channel.

$sel:fromAddress:EmailChannelRequest', emailChannelRequest_fromAddress - The verified email address that you want to send email from when you send email through the channel.

$sel:identity:EmailChannelRequest', emailChannelRequest_identity - The Amazon Resource Name (ARN) of the identity, verified with Amazon Simple Email Service (Amazon SES), that you want to use when you send email through the channel.

emailChannelRequest_enabled :: Lens' EmailChannelRequest (Maybe Bool)

Specifies whether to enable the email channel for the application.

emailChannelRequest_configurationSet :: Lens' EmailChannelRequest (Maybe Text)

The Amazon SES configuration set that you want to apply to messages that you send through the channel.

emailChannelRequest_roleArn :: Lens' EmailChannelRequest (Maybe Text)

The ARN of the AWS Identity and Access Management (IAM) role that you want Amazon Pinpoint to use when it submits email-related event data for the channel.

emailChannelRequest_fromAddress :: Lens' EmailChannelRequest Text

The verified email address that you want to send email from when you send email through the channel.

emailChannelRequest_identity :: Lens' EmailChannelRequest Text

The Amazon Resource Name (ARN) of the identity, verified with Amazon Simple Email Service (Amazon SES), that you want to use when you send email through the channel.