Network.AWS.Pinpoint.Types.EmailChannelResponse

Description

Documentation

data EmailChannelResponse

Provides information about the status and settings of the email channel for an application.

See: newEmailChannelResponse smart constructor.

Constructors

EmailChannelResponse'

Fields

applicationId :: Maybe Text

The unique identifier for the application that the email channel applies to.

lastModifiedDate :: Maybe Text

The date and time, in ISO 8601 format, when the email channel was last modified.

hasCredential :: Maybe Bool

(Not used) This property is retained only for backward compatibility.

roleArn :: Maybe Text

The ARN of the AWS Identity and Access Management (IAM) role that Amazon Pinpoint uses to submit email-related event data for the channel.

identity :: Maybe Text

The Amazon Resource Name (ARN) of the identity, verified with Amazon Simple Email Service (Amazon SES), that's used when you send email through the channel.

enabled :: Maybe Bool

Specifies whether the email channel is enabled for the application.

creationDate :: Maybe Text

The date and time, in ISO 8601 format, when the email channel was enabled.

id :: Maybe Text

(Deprecated) An identifier for the email channel. This property is retained only for backward compatibility.

version :: Maybe Int

The current version of the email channel.

messagesPerSecond :: Maybe Int

The maximum number of emails that can be sent through the channel each second.

isArchived :: Maybe Bool

Specifies whether the email channel is archived.

fromAddress :: Maybe Text

The verified email address that email is sent from when you send email through the channel.

lastModifiedBy :: Maybe Text

The user who last modified the email channel.

configurationSet :: Maybe Text

The Amazon SES configuration set that's applied to messages that are sent through the channel.

platform :: Text

The type of messaging or notification platform for the channel. For the email channel, this value is EMAIL.

newEmailChannelResponse

Arguments

:: Text
$sel:platform:EmailChannelResponse'

-> EmailChannelResponse

Create a value of EmailChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:EmailChannelResponse', emailChannelResponse_applicationId - The unique identifier for the application that the email channel applies to.

$sel:lastModifiedDate:EmailChannelResponse', emailChannelResponse_lastModifiedDate - The date and time, in ISO 8601 format, when the email channel was last modified.

$sel:hasCredential:EmailChannelResponse', emailChannelResponse_hasCredential - (Not used) This property is retained only for backward compatibility.

$sel:roleArn:EmailChannelResponse', emailChannelResponse_roleArn - The ARN of the AWS Identity and Access Management (IAM) role that Amazon Pinpoint uses to submit email-related event data for the channel.

$sel:identity:EmailChannelResponse', emailChannelResponse_identity - The Amazon Resource Name (ARN) of the identity, verified with Amazon Simple Email Service (Amazon SES), that's used when you send email through the channel.

$sel:enabled:EmailChannelResponse', emailChannelResponse_enabled - Specifies whether the email channel is enabled for the application.

$sel:creationDate:EmailChannelResponse', emailChannelResponse_creationDate - The date and time, in ISO 8601 format, when the email channel was enabled.

$sel:id:EmailChannelResponse', emailChannelResponse_id - (Deprecated) An identifier for the email channel. This property is retained only for backward compatibility.

$sel:version:EmailChannelResponse', emailChannelResponse_version - The current version of the email channel.

$sel:messagesPerSecond:EmailChannelResponse', emailChannelResponse_messagesPerSecond - The maximum number of emails that can be sent through the channel each second.

$sel:isArchived:EmailChannelResponse', emailChannelResponse_isArchived - Specifies whether the email channel is archived.

$sel:fromAddress:EmailChannelResponse', emailChannelResponse_fromAddress - The verified email address that email is sent from when you send email through the channel.

$sel:lastModifiedBy:EmailChannelResponse', emailChannelResponse_lastModifiedBy - The user who last modified the email channel.

$sel:configurationSet:EmailChannelResponse', emailChannelResponse_configurationSet - The Amazon SES configuration set that's applied to messages that are sent through the channel.

$sel:platform:EmailChannelResponse', emailChannelResponse_platform - The type of messaging or notification platform for the channel. For the email channel, this value is EMAIL.

emailChannelResponse_applicationId :: Lens' EmailChannelResponse (Maybe Text)

The unique identifier for the application that the email channel applies to.

emailChannelResponse_lastModifiedDate :: Lens' EmailChannelResponse (Maybe Text)

The date and time, in ISO 8601 format, when the email channel was last modified.

emailChannelResponse_hasCredential :: Lens' EmailChannelResponse (Maybe Bool)

(Not used) This property is retained only for backward compatibility.

emailChannelResponse_roleArn :: Lens' EmailChannelResponse (Maybe Text)

The ARN of the AWS Identity and Access Management (IAM) role that Amazon Pinpoint uses to submit email-related event data for the channel.

emailChannelResponse_identity :: Lens' EmailChannelResponse (Maybe Text)

The Amazon Resource Name (ARN) of the identity, verified with Amazon Simple Email Service (Amazon SES), that's used when you send email through the channel.

emailChannelResponse_enabled :: Lens' EmailChannelResponse (Maybe Bool)

Specifies whether the email channel is enabled for the application.

emailChannelResponse_creationDate :: Lens' EmailChannelResponse (Maybe Text)

The date and time, in ISO 8601 format, when the email channel was enabled.

emailChannelResponse_id :: Lens' EmailChannelResponse (Maybe Text)

(Deprecated) An identifier for the email channel. This property is retained only for backward compatibility.

emailChannelResponse_version :: Lens' EmailChannelResponse (Maybe Int)

The current version of the email channel.

emailChannelResponse_messagesPerSecond :: Lens' EmailChannelResponse (Maybe Int)

The maximum number of emails that can be sent through the channel each second.

emailChannelResponse_isArchived :: Lens' EmailChannelResponse (Maybe Bool)

Specifies whether the email channel is archived.

emailChannelResponse_fromAddress :: Lens' EmailChannelResponse (Maybe Text)

The verified email address that email is sent from when you send email through the channel.

emailChannelResponse_lastModifiedBy :: Lens' EmailChannelResponse (Maybe Text)

The user who last modified the email channel.

emailChannelResponse_configurationSet :: Lens' EmailChannelResponse (Maybe Text)

The Amazon SES configuration set that's applied to messages that are sent through the channel.

emailChannelResponse_platform :: Lens' EmailChannelResponse Text

The type of messaging or notification platform for the channel. For the email channel, this value is EMAIL.