Network.AWS.Pinpoint.Types.EmailTemplateRequest

Description

Documentation

data EmailTemplateRequest

Specifies the content and settings for a message template that can be used in messages that are sent through the email channel.

See: newEmailTemplateRequest smart constructor.

Constructors

EmailTemplateRequest'

Fields

templateDescription :: Maybe Text

A custom description of the message template.

textPart :: Maybe Text

The message body, in plain text format, to use in email messages that are based on the message template. We recommend using plain text format for email clients that don't render HTML content and clients that are connected to high-latency networks, such as mobile devices.

defaultSubstitutions :: Maybe Text

A JSON object that specifies the default values to use for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable. When you create a message that's based on the template, you can override these defaults with message-specific and address-specific variables and values.

subject :: Maybe Text

The subject line, or title, to use in email messages that are based on the message template.

tags :: Maybe (HashMap Text Text)

A string-to-string map of key-value pairs that defines the tags to associate with the message template. Each tag consists of a required tag key and an associated tag value.

recommenderId :: Maybe Text

The unique identifier for the recommender model to use for the message template. Amazon Pinpoint uses this value to determine how to retrieve and process data from a recommender model when it sends messages that use the template, if the template contains message variables for recommendation data.

htmlPart :: Maybe Text

The message body, in HTML format, to use in email messages that are based on the message template. We recommend using HTML format for email clients that render HTML content. You can include links, formatted text, and more in an HTML message.

newEmailTemplateRequest :: EmailTemplateRequest

Create a value of EmailTemplateRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateDescription:EmailTemplateRequest', emailTemplateRequest_templateDescription - A custom description of the message template.

$sel:textPart:EmailTemplateRequest', emailTemplateRequest_textPart - The message body, in plain text format, to use in email messages that are based on the message template. We recommend using plain text format for email clients that don't render HTML content and clients that are connected to high-latency networks, such as mobile devices.

$sel:defaultSubstitutions:EmailTemplateRequest', emailTemplateRequest_defaultSubstitutions - A JSON object that specifies the default values to use for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable. When you create a message that's based on the template, you can override these defaults with message-specific and address-specific variables and values.

$sel:subject:EmailTemplateRequest', emailTemplateRequest_subject - The subject line, or title, to use in email messages that are based on the message template.

$sel:tags:EmailTemplateRequest', emailTemplateRequest_tags - A string-to-string map of key-value pairs that defines the tags to associate with the message template. Each tag consists of a required tag key and an associated tag value.

$sel:recommenderId:EmailTemplateRequest', emailTemplateRequest_recommenderId - The unique identifier for the recommender model to use for the message template. Amazon Pinpoint uses this value to determine how to retrieve and process data from a recommender model when it sends messages that use the template, if the template contains message variables for recommendation data.

$sel:htmlPart:EmailTemplateRequest', emailTemplateRequest_htmlPart - The message body, in HTML format, to use in email messages that are based on the message template. We recommend using HTML format for email clients that render HTML content. You can include links, formatted text, and more in an HTML message.

emailTemplateRequest_templateDescription :: Lens' EmailTemplateRequest (Maybe Text)

A custom description of the message template.

emailTemplateRequest_textPart :: Lens' EmailTemplateRequest (Maybe Text)

The message body, in plain text format, to use in email messages that are based on the message template. We recommend using plain text format for email clients that don't render HTML content and clients that are connected to high-latency networks, such as mobile devices.

emailTemplateRequest_defaultSubstitutions :: Lens' EmailTemplateRequest (Maybe Text)

A JSON object that specifies the default values to use for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable. When you create a message that's based on the template, you can override these defaults with message-specific and address-specific variables and values.

emailTemplateRequest_subject :: Lens' EmailTemplateRequest (Maybe Text)

The subject line, or title, to use in email messages that are based on the message template.

emailTemplateRequest_tags :: Lens' EmailTemplateRequest (Maybe (HashMap Text Text))

A string-to-string map of key-value pairs that defines the tags to associate with the message template. Each tag consists of a required tag key and an associated tag value.

emailTemplateRequest_recommenderId :: Lens' EmailTemplateRequest (Maybe Text)

The unique identifier for the recommender model to use for the message template. Amazon Pinpoint uses this value to determine how to retrieve and process data from a recommender model when it sends messages that use the template, if the template contains message variables for recommendation data.

emailTemplateRequest_htmlPart :: Lens' EmailTemplateRequest (Maybe Text)

The message body, in HTML format, to use in email messages that are based on the message template. We recommend using HTML format for email clients that render HTML content. You can include links, formatted text, and more in an HTML message.