Network.AWS.Pinpoint.Types.EndpointBatchItem

Description

Documentation

data EndpointBatchItem

Specifies an endpoint to create or update and the settings and attributes to set or change for the endpoint.

See: newEndpointBatchItem smart constructor.

Constructors

EndpointBatchItem'

Fields

user :: Maybe EndpointUser

One or more custom attributes that describe the user who's associated with the endpoint.

address :: Maybe Text

The destination address for messages or push notifications that you send to the endpoint. The address varies by channel. For a push-notification channel, use the token provided by the push notification service, such as an Apple Push Notification service (APNs) device token or a Firebase Cloud Messaging (FCM) registration token. For the SMS channel, use a phone number in E.164 format, such as +12065550100. For the email channel, use an email address.

channelType :: Maybe ChannelType

The channel to use when sending messages or push notifications to the endpoint.

id :: Maybe Text

The unique identifier for the endpoint in the context of the batch.

optOut :: Maybe Text

Specifies whether the user who's associated with the endpoint has opted out of receiving messages and push notifications from you. Possible values are: ALL, the user has opted out and doesn't want to receive any messages or push notifications; and, NONE, the user hasn't opted out and wants to receive all messages and push notifications.

demographic :: Maybe EndpointDemographic

The demographic information for the endpoint, such as the time zone and platform.

attributes :: Maybe (HashMap Text [Text])

One or more custom attributes that describe the endpoint by associating a name with an array of values. For example, the value of a custom attribute named Interests might be: ["Science", "Music", "Travel"]. You can use these attributes as filter criteria when you create segments. Attribute names are case sensitive.

An attribute name can contain up to 50 characters. An attribute value can contain up to 100 characters. When you define the name of a custom attribute, avoid using the following characters: number sign (#), colon (:), question mark (?), backslash (\), and slash (/). The Amazon Pinpoint console can't display attribute names that contain these characters. This restriction doesn't apply to attribute values.

metrics :: Maybe (HashMap Text Double)

One or more custom metrics that your app reports to Amazon Pinpoint for the endpoint.

endpointStatus :: Maybe Text

Specifies whether to send messages or push notifications to the endpoint. Valid values are: ACTIVE, messages are sent to the endpoint; and, INACTIVE, messages aren’t sent to the endpoint.

Amazon Pinpoint automatically sets this value to ACTIVE when you create an endpoint or update an existing endpoint. Amazon Pinpoint automatically sets this value to INACTIVE if you update another endpoint that has the same address specified by the Address property.

requestId :: Maybe Text

The unique identifier for the request to create or update the endpoint.

effectiveDate :: Maybe Text

The date and time, in ISO 8601 format, when the endpoint was created or updated.

location :: Maybe EndpointLocation

The geographic information for the endpoint.

newEndpointBatchItem :: EndpointBatchItem

Create a value of EndpointBatchItem with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:user:EndpointBatchItem', endpointBatchItem_user - One or more custom attributes that describe the user who's associated with the endpoint.

$sel:address:EndpointBatchItem', endpointBatchItem_address - The destination address for messages or push notifications that you send to the endpoint. The address varies by channel. For a push-notification channel, use the token provided by the push notification service, such as an Apple Push Notification service (APNs) device token or a Firebase Cloud Messaging (FCM) registration token. For the SMS channel, use a phone number in E.164 format, such as +12065550100. For the email channel, use an email address.

$sel:channelType:EndpointBatchItem', endpointBatchItem_channelType - The channel to use when sending messages or push notifications to the endpoint.

$sel:id:EndpointBatchItem', endpointBatchItem_id - The unique identifier for the endpoint in the context of the batch.

$sel:optOut:EndpointBatchItem', endpointBatchItem_optOut - Specifies whether the user who's associated with the endpoint has opted out of receiving messages and push notifications from you. Possible values are: ALL, the user has opted out and doesn't want to receive any messages or push notifications; and, NONE, the user hasn't opted out and wants to receive all messages and push notifications.

$sel:demographic:EndpointBatchItem', endpointBatchItem_demographic - The demographic information for the endpoint, such as the time zone and platform.

$sel:attributes:EndpointBatchItem', endpointBatchItem_attributes - One or more custom attributes that describe the endpoint by associating a name with an array of values. For example, the value of a custom attribute named Interests might be: ["Science", "Music", "Travel"]. You can use these attributes as filter criteria when you create segments. Attribute names are case sensitive.

An attribute name can contain up to 50 characters. An attribute value can contain up to 100 characters. When you define the name of a custom attribute, avoid using the following characters: number sign (#), colon (:), question mark (?), backslash (\), and slash (/). The Amazon Pinpoint console can't display attribute names that contain these characters. This restriction doesn't apply to attribute values.

$sel:metrics:EndpointBatchItem', endpointBatchItem_metrics - One or more custom metrics that your app reports to Amazon Pinpoint for the endpoint.

$sel:endpointStatus:EndpointBatchItem', endpointBatchItem_endpointStatus - Specifies whether to send messages or push notifications to the endpoint. Valid values are: ACTIVE, messages are sent to the endpoint; and, INACTIVE, messages aren’t sent to the endpoint.

Amazon Pinpoint automatically sets this value to ACTIVE when you create an endpoint or update an existing endpoint. Amazon Pinpoint automatically sets this value to INACTIVE if you update another endpoint that has the same address specified by the Address property.

$sel:requestId:EndpointBatchItem', endpointBatchItem_requestId - The unique identifier for the request to create or update the endpoint.

$sel:effectiveDate:EndpointBatchItem', endpointBatchItem_effectiveDate - The date and time, in ISO 8601 format, when the endpoint was created or updated.

$sel:location:EndpointBatchItem', endpointBatchItem_location - The geographic information for the endpoint.

endpointBatchItem_user :: Lens' EndpointBatchItem (Maybe EndpointUser)

One or more custom attributes that describe the user who's associated with the endpoint.

endpointBatchItem_address :: Lens' EndpointBatchItem (Maybe Text)

The destination address for messages or push notifications that you send to the endpoint. The address varies by channel. For a push-notification channel, use the token provided by the push notification service, such as an Apple Push Notification service (APNs) device token or a Firebase Cloud Messaging (FCM) registration token. For the SMS channel, use a phone number in E.164 format, such as +12065550100. For the email channel, use an email address.

endpointBatchItem_channelType :: Lens' EndpointBatchItem (Maybe ChannelType)

The channel to use when sending messages or push notifications to the endpoint.

endpointBatchItem_id :: Lens' EndpointBatchItem (Maybe Text)

The unique identifier for the endpoint in the context of the batch.

endpointBatchItem_optOut :: Lens' EndpointBatchItem (Maybe Text)

Specifies whether the user who's associated with the endpoint has opted out of receiving messages and push notifications from you. Possible values are: ALL, the user has opted out and doesn't want to receive any messages or push notifications; and, NONE, the user hasn't opted out and wants to receive all messages and push notifications.

endpointBatchItem_demographic :: Lens' EndpointBatchItem (Maybe EndpointDemographic)

The demographic information for the endpoint, such as the time zone and platform.

endpointBatchItem_attributes :: Lens' EndpointBatchItem (Maybe (HashMap Text [Text]))

One or more custom attributes that describe the endpoint by associating a name with an array of values. For example, the value of a custom attribute named Interests might be: ["Science", "Music", "Travel"]. You can use these attributes as filter criteria when you create segments. Attribute names are case sensitive.

An attribute name can contain up to 50 characters. An attribute value can contain up to 100 characters. When you define the name of a custom attribute, avoid using the following characters: number sign (#), colon (:), question mark (?), backslash (\), and slash (/). The Amazon Pinpoint console can't display attribute names that contain these characters. This restriction doesn't apply to attribute values.

endpointBatchItem_metrics :: Lens' EndpointBatchItem (Maybe (HashMap Text Double))

One or more custom metrics that your app reports to Amazon Pinpoint for the endpoint.

endpointBatchItem_endpointStatus :: Lens' EndpointBatchItem (Maybe Text)

Specifies whether to send messages or push notifications to the endpoint. Valid values are: ACTIVE, messages are sent to the endpoint; and, INACTIVE, messages aren’t sent to the endpoint.

Amazon Pinpoint automatically sets this value to ACTIVE when you create an endpoint or update an existing endpoint. Amazon Pinpoint automatically sets this value to INACTIVE if you update another endpoint that has the same address specified by the Address property.

endpointBatchItem_requestId :: Lens' EndpointBatchItem (Maybe Text)

The unique identifier for the request to create or update the endpoint.

endpointBatchItem_effectiveDate :: Lens' EndpointBatchItem (Maybe Text)

The date and time, in ISO 8601 format, when the endpoint was created or updated.

endpointBatchItem_location :: Lens' EndpointBatchItem (Maybe EndpointLocation)

The geographic information for the endpoint.