Network.AWS.Pinpoint.Types.EndpointDemographic

Description

Documentation

data EndpointDemographic

Specifies demographic information about an endpoint, such as the applicable time zone and platform.

See: newEndpointDemographic smart constructor.

Constructors

EndpointDemographic'

Fields

model :: Maybe Text

The model name or number of the endpoint device, such as iPhone or SM-G900F.

platform :: Maybe Text

The platform of the endpoint device, such as ios.

appVersion :: Maybe Text

The version of the app that's associated with the endpoint.

locale :: Maybe Text

The locale of the endpoint, in the following format: the ISO 639-1 alpha-2 code, followed by an underscore (_), followed by an ISO 3166-1 alpha-2 value.

modelVersion :: Maybe Text

The model version of the endpoint device.

platformVersion :: Maybe Text

The platform version of the endpoint device.

timezone :: Maybe Text

The time zone of the endpoint, specified as a tz database name value, such as America/Los_Angeles.

make :: Maybe Text

The manufacturer of the endpoint device, such as apple or samsung.

newEndpointDemographic :: EndpointDemographic

Create a value of EndpointDemographic with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:model:EndpointDemographic', endpointDemographic_model - The model name or number of the endpoint device, such as iPhone or SM-G900F.

$sel:platform:EndpointDemographic', endpointDemographic_platform - The platform of the endpoint device, such as ios.

$sel:appVersion:EndpointDemographic', endpointDemographic_appVersion - The version of the app that's associated with the endpoint.

$sel:locale:EndpointDemographic', endpointDemographic_locale - The locale of the endpoint, in the following format: the ISO 639-1 alpha-2 code, followed by an underscore (_), followed by an ISO 3166-1 alpha-2 value.

$sel:modelVersion:EndpointDemographic', endpointDemographic_modelVersion - The model version of the endpoint device.

$sel:platformVersion:EndpointDemographic', endpointDemographic_platformVersion - The platform version of the endpoint device.

$sel:timezone:EndpointDemographic', endpointDemographic_timezone - The time zone of the endpoint, specified as a tz database name value, such as America/Los_Angeles.

$sel:make:EndpointDemographic', endpointDemographic_make - The manufacturer of the endpoint device, such as apple or samsung.

endpointDemographic_model :: Lens' EndpointDemographic (Maybe Text)

The model name or number of the endpoint device, such as iPhone or SM-G900F.

endpointDemographic_platform :: Lens' EndpointDemographic (Maybe Text)

The platform of the endpoint device, such as ios.

endpointDemographic_appVersion :: Lens' EndpointDemographic (Maybe Text)

The version of the app that's associated with the endpoint.

endpointDemographic_locale :: Lens' EndpointDemographic (Maybe Text)

The locale of the endpoint, in the following format: the ISO 639-1 alpha-2 code, followed by an underscore (_), followed by an ISO 3166-1 alpha-2 value.

endpointDemographic_modelVersion :: Lens' EndpointDemographic (Maybe Text)

The model version of the endpoint device.

endpointDemographic_platformVersion :: Lens' EndpointDemographic (Maybe Text)

The platform version of the endpoint device.

endpointDemographic_timezone :: Lens' EndpointDemographic (Maybe Text)

The time zone of the endpoint, specified as a tz database name value, such as America/Los_Angeles.

endpointDemographic_make :: Lens' EndpointDemographic (Maybe Text)

The manufacturer of the endpoint device, such as apple or samsung.