Network.AWS.Pinpoint.Types.EndpointSendConfiguration

Description

Documentation

data EndpointSendConfiguration

Specifies the content, including message variables and attributes, to use in a message that's sent directly to an endpoint.

See: newEndpointSendConfiguration smart constructor.

Constructors

EndpointSendConfiguration'

Fields

substitutions :: Maybe (HashMap Text [Text])

A map of the message variables to merge with the variables specified for the default message (DefaultMessage.Substitutions). The variables specified in this map take precedence over all other variables.

titleOverride :: Maybe Text

The title or subject line of the message. If specified, this value overrides the default message title or subject line.

context :: Maybe (HashMap Text Text)

A map of custom attributes to attach to the message for the address. Attribute names are case sensitive.

For a push notification, this payload is added to the data.pinpoint object. For an email or text message, this payload is added to email/SMS delivery receipt event attributes.

rawContent :: Maybe Text

The raw, JSON-formatted string to use as the payload for the message. If specified, this value overrides all other values for the message.

bodyOverride :: Maybe Text

The body of the message. If specified, this value overrides the default message body.

newEndpointSendConfiguration :: EndpointSendConfiguration

Create a value of EndpointSendConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:substitutions:EndpointSendConfiguration', endpointSendConfiguration_substitutions - A map of the message variables to merge with the variables specified for the default message (DefaultMessage.Substitutions). The variables specified in this map take precedence over all other variables.

$sel:titleOverride:EndpointSendConfiguration', endpointSendConfiguration_titleOverride - The title or subject line of the message. If specified, this value overrides the default message title or subject line.

$sel:context:EndpointSendConfiguration', endpointSendConfiguration_context - A map of custom attributes to attach to the message for the address. Attribute names are case sensitive.

For a push notification, this payload is added to the data.pinpoint object. For an email or text message, this payload is added to email/SMS delivery receipt event attributes.

$sel:rawContent:EndpointSendConfiguration', endpointSendConfiguration_rawContent - The raw, JSON-formatted string to use as the payload for the message. If specified, this value overrides all other values for the message.

$sel:bodyOverride:EndpointSendConfiguration', endpointSendConfiguration_bodyOverride - The body of the message. If specified, this value overrides the default message body.

endpointSendConfiguration_substitutions :: Lens' EndpointSendConfiguration (Maybe (HashMap Text [Text]))

A map of the message variables to merge with the variables specified for the default message (DefaultMessage.Substitutions). The variables specified in this map take precedence over all other variables.

endpointSendConfiguration_titleOverride :: Lens' EndpointSendConfiguration (Maybe Text)

The title or subject line of the message. If specified, this value overrides the default message title or subject line.

endpointSendConfiguration_context :: Lens' EndpointSendConfiguration (Maybe (HashMap Text Text))

A map of custom attributes to attach to the message for the address. Attribute names are case sensitive.

For a push notification, this payload is added to the data.pinpoint object. For an email or text message, this payload is added to email/SMS delivery receipt event attributes.

endpointSendConfiguration_rawContent :: Lens' EndpointSendConfiguration (Maybe Text)

The raw, JSON-formatted string to use as the payload for the message. If specified, this value overrides all other values for the message.

endpointSendConfiguration_bodyOverride :: Lens' EndpointSendConfiguration (Maybe Text)

The body of the message. If specified, this value overrides the default message body.