Network.AWS.Pinpoint.Types.Event

Description

Documentation

data Event

Specifies information about an event that reports data to Amazon Pinpoint.

See: newEvent smart constructor.

Constructors

Event'

Fields

clientSdkVersion :: Maybe Text

The version of the SDK that's running on the client device.

metrics :: Maybe (HashMap Text Double)

One or more custom metrics that are associated with the event.

appVersionCode :: Maybe Text

The version number of the app that's recording the event.

appTitle :: Maybe Text

The title of the app that's recording the event.

appPackageName :: Maybe Text

The package name of the app that's recording the event.

attributes :: Maybe (HashMap Text Text)

One or more custom attributes that are associated with the event.

sdkName :: Maybe Text

The name of the SDK that's being used to record the event.

session :: Maybe Session

Information about the session in which the event occurred.

eventType :: Text

The name of the event.

timestamp :: Text

The date and time, in ISO 8601 format, when the event occurred.

newEvent

Arguments

:: Text
$sel:eventType:Event'

-> Text
$sel:timestamp:Event'

-> Event

Create a value of Event with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientSdkVersion:Event', event_clientSdkVersion - The version of the SDK that's running on the client device.

$sel:metrics:Event', event_metrics - One or more custom metrics that are associated with the event.

$sel:appVersionCode:Event', event_appVersionCode - The version number of the app that's recording the event.

$sel:appTitle:Event', event_appTitle - The title of the app that's recording the event.

$sel:appPackageName:Event', event_appPackageName - The package name of the app that's recording the event.

$sel:attributes:Event', event_attributes - One or more custom attributes that are associated with the event.

$sel:sdkName:Event', event_sdkName - The name of the SDK that's being used to record the event.

$sel:session:Event', event_session - Information about the session in which the event occurred.

$sel:eventType:Event', event_eventType - The name of the event.

$sel:timestamp:Event', event_timestamp - The date and time, in ISO 8601 format, when the event occurred.

event_clientSdkVersion :: Lens' Event (Maybe Text)

The version of the SDK that's running on the client device.

event_metrics :: Lens' Event (Maybe (HashMap Text Double))

One or more custom metrics that are associated with the event.

event_appVersionCode :: Lens' Event (Maybe Text)

The version number of the app that's recording the event.

event_appTitle :: Lens' Event (Maybe Text)

The title of the app that's recording the event.

event_appPackageName :: Lens' Event (Maybe Text)

The package name of the app that's recording the event.

event_attributes :: Lens' Event (Maybe (HashMap Text Text))

One or more custom attributes that are associated with the event.

event_sdkName :: Lens' Event (Maybe Text)

The name of the SDK that's being used to record the event.

event_session :: Lens' Event (Maybe Session)

Information about the session in which the event occurred.

event_eventType :: Lens' Event Text

The name of the event.

event_timestamp :: Lens' Event Text

The date and time, in ISO 8601 format, when the event occurred.