Network.AWS.Pinpoint.Types.EventCondition

Description

Documentation

data EventCondition

Specifies the conditions to evaluate for an event that applies to an activity in a journey.

See: newEventCondition smart constructor.

Constructors

EventCondition'

Fields

dimensions :: Maybe EventDimensions

The dimensions for the event filter to use for the activity.

messageActivity :: Maybe Text

The message identifier (message_id) for the message to use when determining whether message events meet the condition.

newEventCondition :: EventCondition

Create a value of EventCondition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dimensions:EventCondition', eventCondition_dimensions - The dimensions for the event filter to use for the activity.

$sel:messageActivity:EventCondition', eventCondition_messageActivity - The message identifier (message_id) for the message to use when determining whether message events meet the condition.

eventCondition_dimensions :: Lens' EventCondition (Maybe EventDimensions)

The dimensions for the event filter to use for the activity.

eventCondition_messageActivity :: Lens' EventCondition (Maybe Text)

The message identifier (message_id) for the message to use when determining whether message events meet the condition.