Network.AWS.Pinpoint.Types.EventDimensions

Description

Documentation

data EventDimensions

Specifies the dimensions for an event filter that determines when a campaign is sent or a journey activity is performed.

See: newEventDimensions smart constructor.

Constructors

EventDimensions'

Fields

eventType :: Maybe SetDimension

The name of the event that causes the campaign to be sent or the journey activity to be performed. This can be a standard event that Amazon Pinpoint generates, such as _email.delivered. For campaigns, this can also be a custom event that's specific to your application. For information about standard events, see Streaming Amazon Pinpoint Events in the Amazon Pinpoint Developer Guide.

attributes :: Maybe (HashMap Text AttributeDimension)

One or more custom attributes that your application reports to Amazon Pinpoint. You can use these attributes as selection criteria when you create an event filter.

metrics :: Maybe (HashMap Text MetricDimension)

One or more custom metrics that your application reports to Amazon Pinpoint. You can use these metrics as selection criteria when you create an event filter.

newEventDimensions :: EventDimensions

Create a value of EventDimensions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventType:EventDimensions', eventDimensions_eventType - The name of the event that causes the campaign to be sent or the journey activity to be performed. This can be a standard event that Amazon Pinpoint generates, such as _email.delivered. For campaigns, this can also be a custom event that's specific to your application. For information about standard events, see Streaming Amazon Pinpoint Events in the Amazon Pinpoint Developer Guide.

$sel:attributes:EventDimensions', eventDimensions_attributes - One or more custom attributes that your application reports to Amazon Pinpoint. You can use these attributes as selection criteria when you create an event filter.

$sel:metrics:EventDimensions', eventDimensions_metrics - One or more custom metrics that your application reports to Amazon Pinpoint. You can use these metrics as selection criteria when you create an event filter.

eventDimensions_eventType :: Lens' EventDimensions (Maybe SetDimension)

The name of the event that causes the campaign to be sent or the journey activity to be performed. This can be a standard event that Amazon Pinpoint generates, such as _email.delivered. For campaigns, this can also be a custom event that's specific to your application. For information about standard events, see Streaming Amazon Pinpoint Events in the Amazon Pinpoint Developer Guide.

eventDimensions_attributes :: Lens' EventDimensions (Maybe (HashMap Text AttributeDimension))

One or more custom attributes that your application reports to Amazon Pinpoint. You can use these attributes as selection criteria when you create an event filter.

eventDimensions_metrics :: Lens' EventDimensions (Maybe (HashMap Text MetricDimension))

One or more custom metrics that your application reports to Amazon Pinpoint. You can use these metrics as selection criteria when you create an event filter.