Network.AWS.Pinpoint.Types.EventFilter

Description

Documentation

data EventFilter

Specifies the settings for an event that causes a campaign to be sent or a journey activity to be performed.

See: newEventFilter smart constructor.

Constructors

EventFilter'

Fields

filterType :: FilterType

The type of event that causes the campaign to be sent or the journey activity to be performed. Valid values are: SYSTEM, sends the campaign or performs the activity when a system event occurs; and, ENDPOINT, sends the campaign or performs the activity when an endpoint event (Events resource) occurs.

dimensions :: EventDimensions

The dimensions for the event filter to use for the campaign or the journey activity.

newEventFilter

Arguments

:: FilterType
$sel:filterType:EventFilter'

-> EventDimensions
$sel:dimensions:EventFilter'

-> EventFilter

Create a value of EventFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filterType:EventFilter', eventFilter_filterType - The type of event that causes the campaign to be sent or the journey activity to be performed. Valid values are: SYSTEM, sends the campaign or performs the activity when a system event occurs; and, ENDPOINT, sends the campaign or performs the activity when an endpoint event (Events resource) occurs.

$sel:dimensions:EventFilter', eventFilter_dimensions - The dimensions for the event filter to use for the campaign or the journey activity.

eventFilter_filterType :: Lens' EventFilter FilterType

The type of event that causes the campaign to be sent or the journey activity to be performed. Valid values are: SYSTEM, sends the campaign or performs the activity when a system event occurs; and, ENDPOINT, sends the campaign or performs the activity when an endpoint event (Events resource) occurs.

eventFilter_dimensions :: Lens' EventFilter EventDimensions

The dimensions for the event filter to use for the campaign or the journey activity.