Network.AWS.Pinpoint.Types.EventItemResponse

Description

Documentation

data EventItemResponse

Provides the status code and message that result from processing an event.

See: newEventItemResponse smart constructor.

Constructors

EventItemResponse'

Fields

message :: Maybe Text

A custom message that's returned in the response as a result of processing the event.

statusCode :: Maybe Int

The status code that's returned in the response as a result of processing the event. Possible values are: 202, for events that were accepted; and, 400, for events that weren't valid.

newEventItemResponse :: EventItemResponse

Create a value of EventItemResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:message:EventItemResponse', eventItemResponse_message - A custom message that's returned in the response as a result of processing the event.

$sel:statusCode:EventItemResponse', eventItemResponse_statusCode - The status code that's returned in the response as a result of processing the event. Possible values are: 202, for events that were accepted; and, 400, for events that weren't valid.

eventItemResponse_message :: Lens' EventItemResponse (Maybe Text)

A custom message that's returned in the response as a result of processing the event.

eventItemResponse_statusCode :: Lens' EventItemResponse (Maybe Int)

The status code that's returned in the response as a result of processing the event. Possible values are: 202, for events that were accepted; and, 400, for events that weren't valid.