Network.AWS.Pinpoint.Types.ExportJobRequest

Description

Documentation

data ExportJobRequest

Specifies the settings for a job that exports endpoint definitions to an Amazon Simple Storage Service (Amazon S3) bucket.

See: newExportJobRequest smart constructor.

Constructors

ExportJobRequest'

Fields

segmentId :: Maybe Text

The identifier for the segment to export endpoint definitions from. If you don't specify this value, Amazon Pinpoint exports definitions for all the endpoints that are associated with the application.

segmentVersion :: Maybe Int

The version of the segment to export endpoint definitions from, if specified.

s3UrlPrefix :: Text

The URL of the location in an Amazon Simple Storage Service (Amazon S3) bucket where you want to export endpoint definitions to. This location is typically a folder that contains multiple files. The URL should be in the following format: s3://bucket-name/folder-name/.

roleArn :: Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that authorizes Amazon Pinpoint to access the Amazon S3 location where you want to export endpoint definitions to.

newExportJobRequest

Arguments

:: Text
$sel:s3UrlPrefix:ExportJobRequest'

-> Text
$sel:roleArn:ExportJobRequest'

-> ExportJobRequest

Create a value of ExportJobRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:segmentId:ExportJobRequest', exportJobRequest_segmentId - The identifier for the segment to export endpoint definitions from. If you don't specify this value, Amazon Pinpoint exports definitions for all the endpoints that are associated with the application.

$sel:segmentVersion:ExportJobRequest', exportJobRequest_segmentVersion - The version of the segment to export endpoint definitions from, if specified.

$sel:s3UrlPrefix:ExportJobRequest', exportJobRequest_s3UrlPrefix - The URL of the location in an Amazon Simple Storage Service (Amazon S3) bucket where you want to export endpoint definitions to. This location is typically a folder that contains multiple files. The URL should be in the following format: s3://bucket-name/folder-name/.

$sel:roleArn:ExportJobRequest', exportJobRequest_roleArn - The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that authorizes Amazon Pinpoint to access the Amazon S3 location where you want to export endpoint definitions to.

exportJobRequest_segmentId :: Lens' ExportJobRequest (Maybe Text)

The identifier for the segment to export endpoint definitions from. If you don't specify this value, Amazon Pinpoint exports definitions for all the endpoints that are associated with the application.

exportJobRequest_segmentVersion :: Lens' ExportJobRequest (Maybe Int)

The version of the segment to export endpoint definitions from, if specified.

exportJobRequest_s3UrlPrefix :: Lens' ExportJobRequest Text

The URL of the location in an Amazon Simple Storage Service (Amazon S3) bucket where you want to export endpoint definitions to. This location is typically a folder that contains multiple files. The URL should be in the following format: s3://bucket-name/folder-name/.

exportJobRequest_roleArn :: Lens' ExportJobRequest Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that authorizes Amazon Pinpoint to access the Amazon S3 location where you want to export endpoint definitions to.