Network.AWS.Pinpoint.Types.ExportJobResponse

Description

Documentation

data ExportJobResponse

Provides information about the status and settings of a job that exports endpoint definitions to a file. The file can be added directly to an Amazon Simple Storage Service (Amazon S3) bucket by using the Amazon Pinpoint API or downloaded directly to a computer by using the Amazon Pinpoint console.

See: newExportJobResponse smart constructor.

Constructors

ExportJobResponse'

Fields

totalFailures :: Maybe Int

The total number of endpoint definitions that weren't processed successfully (failed) by the export job, typically because an error, such as a syntax error, occurred.

failures :: Maybe [Text]

An array of entries, one for each of the first 100 entries that weren't processed successfully (failed) by the export job, if any.

totalProcessed :: Maybe Int

The total number of endpoint definitions that were processed by the export job.

failedPieces :: Maybe Int

The number of pieces that weren't processed successfully (failed) by the export job, as of the time of the request.

completedPieces :: Maybe Int

The number of pieces that were processed successfully (completed) by the export job, as of the time of the request.

totalPieces :: Maybe Int

The total number of pieces that must be processed to complete the export job. Each piece consists of an approximately equal portion of the endpoint definitions that are part of the export job.

completionDate :: Maybe Text

The date, in ISO 8601 format, when the export job was completed.

jobStatus :: JobStatus

The status of the export job. The job status is FAILED if Amazon Pinpoint wasn't able to process one or more pieces in the job.

creationDate :: Text

The date, in ISO 8601 format, when the export job was created.

type' :: Text

The job type. This value is EXPORT for export jobs.

definition :: ExportJobResource

The resource settings that apply to the export job.

id :: Text

The unique identifier for the export job.

applicationId :: Text

The unique identifier for the application that's associated with the export job.

newExportJobResponse

Arguments

:: JobStatus
$sel:jobStatus:ExportJobResponse'

-> Text
$sel:creationDate:ExportJobResponse'

-> Text
$sel:type':ExportJobResponse'

-> ExportJobResource
$sel:definition:ExportJobResponse'

-> Text
$sel:id:ExportJobResponse'

-> Text
$sel:applicationId:ExportJobResponse'

-> ExportJobResponse

Create a value of ExportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:totalFailures:ExportJobResponse', exportJobResponse_totalFailures - The total number of endpoint definitions that weren't processed successfully (failed) by the export job, typically because an error, such as a syntax error, occurred.

$sel:failures:ExportJobResponse', exportJobResponse_failures - An array of entries, one for each of the first 100 entries that weren't processed successfully (failed) by the export job, if any.

$sel:totalProcessed:ExportJobResponse', exportJobResponse_totalProcessed - The total number of endpoint definitions that were processed by the export job.

$sel:failedPieces:ExportJobResponse', exportJobResponse_failedPieces - The number of pieces that weren't processed successfully (failed) by the export job, as of the time of the request.

$sel:completedPieces:ExportJobResponse', exportJobResponse_completedPieces - The number of pieces that were processed successfully (completed) by the export job, as of the time of the request.

$sel:totalPieces:ExportJobResponse', exportJobResponse_totalPieces - The total number of pieces that must be processed to complete the export job. Each piece consists of an approximately equal portion of the endpoint definitions that are part of the export job.

$sel:completionDate:ExportJobResponse', exportJobResponse_completionDate - The date, in ISO 8601 format, when the export job was completed.

$sel:jobStatus:ExportJobResponse', exportJobResponse_jobStatus - The status of the export job. The job status is FAILED if Amazon Pinpoint wasn't able to process one or more pieces in the job.

$sel:creationDate:ExportJobResponse', exportJobResponse_creationDate - The date, in ISO 8601 format, when the export job was created.

$sel:type':ExportJobResponse', exportJobResponse_type - The job type. This value is EXPORT for export jobs.

$sel:definition:ExportJobResponse', exportJobResponse_definition - The resource settings that apply to the export job.

$sel:id:ExportJobResponse', exportJobResponse_id - The unique identifier for the export job.

$sel:applicationId:ExportJobResponse', exportJobResponse_applicationId - The unique identifier for the application that's associated with the export job.

exportJobResponse_totalFailures :: Lens' ExportJobResponse (Maybe Int)

The total number of endpoint definitions that weren't processed successfully (failed) by the export job, typically because an error, such as a syntax error, occurred.

exportJobResponse_failures :: Lens' ExportJobResponse (Maybe [Text])

An array of entries, one for each of the first 100 entries that weren't processed successfully (failed) by the export job, if any.

exportJobResponse_totalProcessed :: Lens' ExportJobResponse (Maybe Int)

The total number of endpoint definitions that were processed by the export job.

exportJobResponse_failedPieces :: Lens' ExportJobResponse (Maybe Int)

The number of pieces that weren't processed successfully (failed) by the export job, as of the time of the request.

exportJobResponse_completedPieces :: Lens' ExportJobResponse (Maybe Int)

The number of pieces that were processed successfully (completed) by the export job, as of the time of the request.

exportJobResponse_totalPieces :: Lens' ExportJobResponse (Maybe Int)

The total number of pieces that must be processed to complete the export job. Each piece consists of an approximately equal portion of the endpoint definitions that are part of the export job.

exportJobResponse_completionDate :: Lens' ExportJobResponse (Maybe Text)

The date, in ISO 8601 format, when the export job was completed.

exportJobResponse_jobStatus :: Lens' ExportJobResponse JobStatus

The status of the export job. The job status is FAILED if Amazon Pinpoint wasn't able to process one or more pieces in the job.

exportJobResponse_creationDate :: Lens' ExportJobResponse Text

The date, in ISO 8601 format, when the export job was created.

exportJobResponse_type :: Lens' ExportJobResponse Text

The job type. This value is EXPORT for export jobs.

exportJobResponse_definition :: Lens' ExportJobResponse ExportJobResource

The resource settings that apply to the export job.

exportJobResponse_id :: Lens' ExportJobResponse Text

The unique identifier for the export job.

exportJobResponse_applicationId :: Lens' ExportJobResponse Text

The unique identifier for the application that's associated with the export job.