Network.AWS.Pinpoint.Types.GCMChannelResponse

Description

Documentation

data GCMChannelResponse

Provides information about the status and settings of the GCM channel for an application. The GCM channel enables Amazon Pinpoint to send push notifications through the Firebase Cloud Messaging (FCM), formerly Google Cloud Messaging (GCM), service.

See: newGCMChannelResponse smart constructor.

Constructors

GCMChannelResponse'

Fields

applicationId :: Maybe Text

The unique identifier for the application that the GCM channel applies to.

lastModifiedDate :: Maybe Text

The date and time when the GCM channel was last modified.

hasCredential :: Maybe Bool

(Not used) This property is retained only for backward compatibility.

enabled :: Maybe Bool

Specifies whether the GCM channel is enabled for the application.

creationDate :: Maybe Text

The date and time when the GCM channel was enabled.

id :: Maybe Text

(Deprecated) An identifier for the GCM channel. This property is retained only for backward compatibility.

version :: Maybe Int

The current version of the GCM channel.

isArchived :: Maybe Bool

Specifies whether the GCM channel is archived.

lastModifiedBy :: Maybe Text

The user who last modified the GCM channel.

credential :: Text

The Web API Key, also referred to as an API_KEY or server key, that you received from Google to communicate with Google services.

platform :: Text

The type of messaging or notification platform for the channel. For the GCM channel, this value is GCM.

newGCMChannelResponse

Arguments

:: Text
$sel:credential:GCMChannelResponse'

-> Text
$sel:platform:GCMChannelResponse'

-> GCMChannelResponse

Create a value of GCMChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applicationId:GCMChannelResponse', gCMChannelResponse_applicationId - The unique identifier for the application that the GCM channel applies to.

$sel:lastModifiedDate:GCMChannelResponse', gCMChannelResponse_lastModifiedDate - The date and time when the GCM channel was last modified.

$sel:hasCredential:GCMChannelResponse', gCMChannelResponse_hasCredential - (Not used) This property is retained only for backward compatibility.

$sel:enabled:GCMChannelResponse', gCMChannelResponse_enabled - Specifies whether the GCM channel is enabled for the application.

$sel:creationDate:GCMChannelResponse', gCMChannelResponse_creationDate - The date and time when the GCM channel was enabled.

$sel:id:GCMChannelResponse', gCMChannelResponse_id - (Deprecated) An identifier for the GCM channel. This property is retained only for backward compatibility.

$sel:version:GCMChannelResponse', gCMChannelResponse_version - The current version of the GCM channel.

$sel:isArchived:GCMChannelResponse', gCMChannelResponse_isArchived - Specifies whether the GCM channel is archived.

$sel:lastModifiedBy:GCMChannelResponse', gCMChannelResponse_lastModifiedBy - The user who last modified the GCM channel.

$sel:credential:GCMChannelResponse', gCMChannelResponse_credential - The Web API Key, also referred to as an API_KEY or server key, that you received from Google to communicate with Google services.

$sel:platform:GCMChannelResponse', gCMChannelResponse_platform - The type of messaging or notification platform for the channel. For the GCM channel, this value is GCM.

gCMChannelResponse_applicationId :: Lens' GCMChannelResponse (Maybe Text)

The unique identifier for the application that the GCM channel applies to.

gCMChannelResponse_lastModifiedDate :: Lens' GCMChannelResponse (Maybe Text)

The date and time when the GCM channel was last modified.

gCMChannelResponse_hasCredential :: Lens' GCMChannelResponse (Maybe Bool)

(Not used) This property is retained only for backward compatibility.

gCMChannelResponse_enabled :: Lens' GCMChannelResponse (Maybe Bool)

Specifies whether the GCM channel is enabled for the application.

gCMChannelResponse_creationDate :: Lens' GCMChannelResponse (Maybe Text)

The date and time when the GCM channel was enabled.

gCMChannelResponse_id :: Lens' GCMChannelResponse (Maybe Text)

(Deprecated) An identifier for the GCM channel. This property is retained only for backward compatibility.

gCMChannelResponse_version :: Lens' GCMChannelResponse (Maybe Int)

The current version of the GCM channel.

gCMChannelResponse_isArchived :: Lens' GCMChannelResponse (Maybe Bool)

Specifies whether the GCM channel is archived.

gCMChannelResponse_lastModifiedBy :: Lens' GCMChannelResponse (Maybe Text)

The user who last modified the GCM channel.

gCMChannelResponse_credential :: Lens' GCMChannelResponse Text

The Web API Key, also referred to as an API_KEY or server key, that you received from Google to communicate with Google services.

gCMChannelResponse_platform :: Lens' GCMChannelResponse Text

The type of messaging or notification platform for the channel. For the GCM channel, this value is GCM.