Network.AWS.Pinpoint.Types.ImportJobRequest

Description

Documentation

data ImportJobRequest

Specifies the settings for a job that imports endpoint definitions from an Amazon Simple Storage Service (Amazon S3) bucket.

See: newImportJobRequest smart constructor.

Constructors

ImportJobRequest'

Fields

segmentName :: Maybe Text

A custom name for the segment that's created by the import job, if the value of the DefineSegment property is true.

defineSegment :: Maybe Bool

Specifies whether to create a segment that contains the endpoints, when the endpoint definitions are imported.

registerEndpoints :: Maybe Bool

Specifies whether to register the endpoints with Amazon Pinpoint, when the endpoint definitions are imported.

externalId :: Maybe Text

(Deprecated) Your AWS account ID, which you assigned to an external ID key in an IAM trust policy. Amazon Pinpoint previously used this value to assume an IAM role when importing endpoint definitions, but we removed this requirement. We don't recommend use of external IDs for IAM roles that are assumed by Amazon Pinpoint.

segmentId :: Maybe Text

The identifier for the segment to update or add the imported endpoint definitions to, if the import job is meant to update an existing segment.

format :: DefinitionFormat

The format of the files that contain the endpoint definitions to import. Valid values are: CSV, for comma-separated values format; and, JSON, for newline-delimited JSON format. If the Amazon S3 location stores multiple files that use different formats, Amazon Pinpoint imports data only from the files that use the specified format.

s3Url :: Text

The URL of the Amazon Simple Storage Service (Amazon S3) bucket that contains the endpoint definitions to import. This location can be a folder or a single file. If the location is a folder, Amazon Pinpoint imports endpoint definitions from the files in this location, including any subfolders that the folder contains.

The URL should be in the following format: s3://bucket-name/folder-name/file-name. The location can end with the key for an individual object or a prefix that qualifies multiple objects.

roleArn :: Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that authorizes Amazon Pinpoint to access the Amazon S3 location to import endpoint definitions from.

newImportJobRequest

Arguments

:: DefinitionFormat
$sel:format:ImportJobRequest'

-> Text
$sel:s3Url:ImportJobRequest'

-> Text
$sel:roleArn:ImportJobRequest'

-> ImportJobRequest

Create a value of ImportJobRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:segmentName:ImportJobRequest', importJobRequest_segmentName - A custom name for the segment that's created by the import job, if the value of the DefineSegment property is true.

$sel:defineSegment:ImportJobRequest', importJobRequest_defineSegment - Specifies whether to create a segment that contains the endpoints, when the endpoint definitions are imported.

$sel:registerEndpoints:ImportJobRequest', importJobRequest_registerEndpoints - Specifies whether to register the endpoints with Amazon Pinpoint, when the endpoint definitions are imported.

$sel:externalId:ImportJobRequest', importJobRequest_externalId - (Deprecated) Your AWS account ID, which you assigned to an external ID key in an IAM trust policy. Amazon Pinpoint previously used this value to assume an IAM role when importing endpoint definitions, but we removed this requirement. We don't recommend use of external IDs for IAM roles that are assumed by Amazon Pinpoint.

$sel:segmentId:ImportJobRequest', importJobRequest_segmentId - The identifier for the segment to update or add the imported endpoint definitions to, if the import job is meant to update an existing segment.

$sel:format:ImportJobRequest', importJobRequest_format - The format of the files that contain the endpoint definitions to import. Valid values are: CSV, for comma-separated values format; and, JSON, for newline-delimited JSON format. If the Amazon S3 location stores multiple files that use different formats, Amazon Pinpoint imports data only from the files that use the specified format.

$sel:s3Url:ImportJobRequest', importJobRequest_s3Url - The URL of the Amazon Simple Storage Service (Amazon S3) bucket that contains the endpoint definitions to import. This location can be a folder or a single file. If the location is a folder, Amazon Pinpoint imports endpoint definitions from the files in this location, including any subfolders that the folder contains.

The URL should be in the following format: s3://bucket-name/folder-name/file-name. The location can end with the key for an individual object or a prefix that qualifies multiple objects.

$sel:roleArn:ImportJobRequest', importJobRequest_roleArn - The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that authorizes Amazon Pinpoint to access the Amazon S3 location to import endpoint definitions from.

importJobRequest_segmentName :: Lens' ImportJobRequest (Maybe Text)

A custom name for the segment that's created by the import job, if the value of the DefineSegment property is true.

importJobRequest_defineSegment :: Lens' ImportJobRequest (Maybe Bool)

Specifies whether to create a segment that contains the endpoints, when the endpoint definitions are imported.

importJobRequest_registerEndpoints :: Lens' ImportJobRequest (Maybe Bool)

Specifies whether to register the endpoints with Amazon Pinpoint, when the endpoint definitions are imported.

importJobRequest_externalId :: Lens' ImportJobRequest (Maybe Text)

(Deprecated) Your AWS account ID, which you assigned to an external ID key in an IAM trust policy. Amazon Pinpoint previously used this value to assume an IAM role when importing endpoint definitions, but we removed this requirement. We don't recommend use of external IDs for IAM roles that are assumed by Amazon Pinpoint.

importJobRequest_segmentId :: Lens' ImportJobRequest (Maybe Text)

The identifier for the segment to update or add the imported endpoint definitions to, if the import job is meant to update an existing segment.

importJobRequest_format :: Lens' ImportJobRequest DefinitionFormat

The format of the files that contain the endpoint definitions to import. Valid values are: CSV, for comma-separated values format; and, JSON, for newline-delimited JSON format. If the Amazon S3 location stores multiple files that use different formats, Amazon Pinpoint imports data only from the files that use the specified format.

importJobRequest_s3Url :: Lens' ImportJobRequest Text

The URL of the Amazon Simple Storage Service (Amazon S3) bucket that contains the endpoint definitions to import. This location can be a folder or a single file. If the location is a folder, Amazon Pinpoint imports endpoint definitions from the files in this location, including any subfolders that the folder contains.

The URL should be in the following format: s3://bucket-name/folder-name/file-name. The location can end with the key for an individual object or a prefix that qualifies multiple objects.

importJobRequest_roleArn :: Lens' ImportJobRequest Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that authorizes Amazon Pinpoint to access the Amazon S3 location to import endpoint definitions from.