Network.AWS.Pinpoint.Types.ImportJobResource

Description

Documentation

data ImportJobResource

Provides information about the resource settings for a job that imports endpoint definitions from one or more files. The files can be stored in an Amazon Simple Storage Service (Amazon S3) bucket or uploaded directly from a computer by using the Amazon Pinpoint console.

See: newImportJobResource smart constructor.

Constructors

ImportJobResource'

Fields

defineSegment :: Maybe Bool

Specifies whether the import job creates a segment that contains the endpoints, when the endpoint definitions are imported.

segmentName :: Maybe Text

The custom name for the segment that's created by the import job, if the value of the DefineSegment property is true.

registerEndpoints :: Maybe Bool

Specifies whether the import job registers the endpoints with Amazon Pinpoint, when the endpoint definitions are imported.

segmentId :: Maybe Text

The identifier for the segment that the import job updates or adds endpoint definitions to, if the import job updates an existing segment.

externalId :: Maybe Text

(Deprecated) Your AWS account ID, which you assigned to an external ID key in an IAM trust policy. Amazon Pinpoint previously used this value to assume an IAM role when importing endpoint definitions, but we removed this requirement. We don't recommend use of external IDs for IAM roles that are assumed by Amazon Pinpoint.

format :: DefinitionFormat

The format of the files that contain the endpoint definitions to import. Valid values are: CSV, for comma-separated values format; and, JSON, for newline-delimited JSON format.

If the files are stored in an Amazon S3 location and that location contains multiple files that use different formats, Amazon Pinpoint imports data only from the files that use the specified format.

s3Url :: Text

The URL of the Amazon Simple Storage Service (Amazon S3) bucket that contains the endpoint definitions to import. This location can be a folder or a single file. If the location is a folder, Amazon Pinpoint imports endpoint definitions from the files in this location, including any subfolders that the folder contains.

The URL should be in the following format: s3://bucket-name/folder-name/file-name. The location can end with the key for an individual object or a prefix that qualifies multiple objects.

roleArn :: Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that authorizes Amazon Pinpoint to access the Amazon S3 location to import endpoint definitions from.

newImportJobResource

Arguments

:: DefinitionFormat
$sel:format:ImportJobResource'

-> Text
$sel:s3Url:ImportJobResource'

-> Text
$sel:roleArn:ImportJobResource'

-> ImportJobResource

Create a value of ImportJobResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defineSegment:ImportJobResource', importJobResource_defineSegment - Specifies whether the import job creates a segment that contains the endpoints, when the endpoint definitions are imported.

$sel:segmentName:ImportJobResource', importJobResource_segmentName - The custom name for the segment that's created by the import job, if the value of the DefineSegment property is true.

$sel:registerEndpoints:ImportJobResource', importJobResource_registerEndpoints - Specifies whether the import job registers the endpoints with Amazon Pinpoint, when the endpoint definitions are imported.

$sel:segmentId:ImportJobResource', importJobResource_segmentId - The identifier for the segment that the import job updates or adds endpoint definitions to, if the import job updates an existing segment.

$sel:externalId:ImportJobResource', importJobResource_externalId - (Deprecated) Your AWS account ID, which you assigned to an external ID key in an IAM trust policy. Amazon Pinpoint previously used this value to assume an IAM role when importing endpoint definitions, but we removed this requirement. We don't recommend use of external IDs for IAM roles that are assumed by Amazon Pinpoint.

$sel:format:ImportJobResource', importJobResource_format - The format of the files that contain the endpoint definitions to import. Valid values are: CSV, for comma-separated values format; and, JSON, for newline-delimited JSON format.

If the files are stored in an Amazon S3 location and that location contains multiple files that use different formats, Amazon Pinpoint imports data only from the files that use the specified format.

$sel:s3Url:ImportJobResource', importJobResource_s3Url - The URL of the Amazon Simple Storage Service (Amazon S3) bucket that contains the endpoint definitions to import. This location can be a folder or a single file. If the location is a folder, Amazon Pinpoint imports endpoint definitions from the files in this location, including any subfolders that the folder contains.

The URL should be in the following format: s3://bucket-name/folder-name/file-name. The location can end with the key for an individual object or a prefix that qualifies multiple objects.

$sel:roleArn:ImportJobResource', importJobResource_roleArn - The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that authorizes Amazon Pinpoint to access the Amazon S3 location to import endpoint definitions from.

importJobResource_defineSegment :: Lens' ImportJobResource (Maybe Bool)

Specifies whether the import job creates a segment that contains the endpoints, when the endpoint definitions are imported.

importJobResource_segmentName :: Lens' ImportJobResource (Maybe Text)

The custom name for the segment that's created by the import job, if the value of the DefineSegment property is true.

importJobResource_registerEndpoints :: Lens' ImportJobResource (Maybe Bool)

Specifies whether the import job registers the endpoints with Amazon Pinpoint, when the endpoint definitions are imported.

importJobResource_segmentId :: Lens' ImportJobResource (Maybe Text)

The identifier for the segment that the import job updates or adds endpoint definitions to, if the import job updates an existing segment.

importJobResource_externalId :: Lens' ImportJobResource (Maybe Text)

(Deprecated) Your AWS account ID, which you assigned to an external ID key in an IAM trust policy. Amazon Pinpoint previously used this value to assume an IAM role when importing endpoint definitions, but we removed this requirement. We don't recommend use of external IDs for IAM roles that are assumed by Amazon Pinpoint.

importJobResource_format :: Lens' ImportJobResource DefinitionFormat

The format of the files that contain the endpoint definitions to import. Valid values are: CSV, for comma-separated values format; and, JSON, for newline-delimited JSON format.

If the files are stored in an Amazon S3 location and that location contains multiple files that use different formats, Amazon Pinpoint imports data only from the files that use the specified format.

importJobResource_s3Url :: Lens' ImportJobResource Text

The URL of the Amazon Simple Storage Service (Amazon S3) bucket that contains the endpoint definitions to import. This location can be a folder or a single file. If the location is a folder, Amazon Pinpoint imports endpoint definitions from the files in this location, including any subfolders that the folder contains.

The URL should be in the following format: s3://bucket-name/folder-name/file-name. The location can end with the key for an individual object or a prefix that qualifies multiple objects.

importJobResource_roleArn :: Lens' ImportJobResource Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that authorizes Amazon Pinpoint to access the Amazon S3 location to import endpoint definitions from.