Network.AWS.Pinpoint.Types.ImportJobResponse

Description

Documentation

data ImportJobResponse

Provides information about the status and settings of a job that imports endpoint definitions from one or more files. The files can be stored in an Amazon Simple Storage Service (Amazon S3) bucket or uploaded directly from a computer by using the Amazon Pinpoint console.

See: newImportJobResponse smart constructor.

Constructors

ImportJobResponse'

Fields

completedPieces :: Maybe Int

The number of pieces that were processed successfully (completed) by the import job, as of the time of the request.

failedPieces :: Maybe Int

The number of pieces that weren't processed successfully (failed) by the import job, as of the time of the request.

totalProcessed :: Maybe Int

The total number of endpoint definitions that were processed by the import job.

failures :: Maybe [Text]

An array of entries, one for each of the first 100 entries that weren't processed successfully (failed) by the import job, if any.

totalPieces :: Maybe Int

The total number of pieces that must be processed to complete the import job. Each piece consists of an approximately equal portion of the endpoint definitions that are part of the import job.

completionDate :: Maybe Text

The date, in ISO 8601 format, when the import job was completed.

totalFailures :: Maybe Int

The total number of endpoint definitions that weren't processed successfully (failed) by the import job, typically because an error, such as a syntax error, occurred.

jobStatus :: JobStatus

The status of the import job. The job status is FAILED if Amazon Pinpoint wasn't able to process one or more pieces in the job.

creationDate :: Text

The date, in ISO 8601 format, when the import job was created.

type' :: Text

The job type. This value is IMPORT for import jobs.

definition :: ImportJobResource

The resource settings that apply to the import job.

id :: Text

The unique identifier for the import job.

applicationId :: Text

The unique identifier for the application that's associated with the import job.

newImportJobResponse

Arguments

:: JobStatus
$sel:jobStatus:ImportJobResponse'

-> Text
$sel:creationDate:ImportJobResponse'

-> Text
$sel:type':ImportJobResponse'

-> ImportJobResource
$sel:definition:ImportJobResponse'

-> Text
$sel:id:ImportJobResponse'

-> Text
$sel:applicationId:ImportJobResponse'

-> ImportJobResponse

Create a value of ImportJobResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:completedPieces:ImportJobResponse', importJobResponse_completedPieces - The number of pieces that were processed successfully (completed) by the import job, as of the time of the request.

$sel:failedPieces:ImportJobResponse', importJobResponse_failedPieces - The number of pieces that weren't processed successfully (failed) by the import job, as of the time of the request.

$sel:totalProcessed:ImportJobResponse', importJobResponse_totalProcessed - The total number of endpoint definitions that were processed by the import job.

$sel:failures:ImportJobResponse', importJobResponse_failures - An array of entries, one for each of the first 100 entries that weren't processed successfully (failed) by the import job, if any.

$sel:totalPieces:ImportJobResponse', importJobResponse_totalPieces - The total number of pieces that must be processed to complete the import job. Each piece consists of an approximately equal portion of the endpoint definitions that are part of the import job.

$sel:completionDate:ImportJobResponse', importJobResponse_completionDate - The date, in ISO 8601 format, when the import job was completed.

$sel:totalFailures:ImportJobResponse', importJobResponse_totalFailures - The total number of endpoint definitions that weren't processed successfully (failed) by the import job, typically because an error, such as a syntax error, occurred.

$sel:jobStatus:ImportJobResponse', importJobResponse_jobStatus - The status of the import job. The job status is FAILED if Amazon Pinpoint wasn't able to process one or more pieces in the job.

$sel:creationDate:ImportJobResponse', importJobResponse_creationDate - The date, in ISO 8601 format, when the import job was created.

$sel:type':ImportJobResponse', importJobResponse_type - The job type. This value is IMPORT for import jobs.

$sel:definition:ImportJobResponse', importJobResponse_definition - The resource settings that apply to the import job.

$sel:id:ImportJobResponse', importJobResponse_id - The unique identifier for the import job.

$sel:applicationId:ImportJobResponse', importJobResponse_applicationId - The unique identifier for the application that's associated with the import job.

importJobResponse_completedPieces :: Lens' ImportJobResponse (Maybe Int)

The number of pieces that were processed successfully (completed) by the import job, as of the time of the request.

importJobResponse_failedPieces :: Lens' ImportJobResponse (Maybe Int)

The number of pieces that weren't processed successfully (failed) by the import job, as of the time of the request.

importJobResponse_totalProcessed :: Lens' ImportJobResponse (Maybe Int)

The total number of endpoint definitions that were processed by the import job.

importJobResponse_failures :: Lens' ImportJobResponse (Maybe [Text])

An array of entries, one for each of the first 100 entries that weren't processed successfully (failed) by the import job, if any.

importJobResponse_totalPieces :: Lens' ImportJobResponse (Maybe Int)

The total number of pieces that must be processed to complete the import job. Each piece consists of an approximately equal portion of the endpoint definitions that are part of the import job.

importJobResponse_completionDate :: Lens' ImportJobResponse (Maybe Text)

The date, in ISO 8601 format, when the import job was completed.

importJobResponse_totalFailures :: Lens' ImportJobResponse (Maybe Int)

The total number of endpoint definitions that weren't processed successfully (failed) by the import job, typically because an error, such as a syntax error, occurred.

importJobResponse_jobStatus :: Lens' ImportJobResponse JobStatus

The status of the import job. The job status is FAILED if Amazon Pinpoint wasn't able to process one or more pieces in the job.

importJobResponse_creationDate :: Lens' ImportJobResponse Text

The date, in ISO 8601 format, when the import job was created.

importJobResponse_type :: Lens' ImportJobResponse Text

The job type. This value is IMPORT for import jobs.

importJobResponse_definition :: Lens' ImportJobResponse ImportJobResource

The resource settings that apply to the import job.

importJobResponse_id :: Lens' ImportJobResponse Text

The unique identifier for the import job.

importJobResponse_applicationId :: Lens' ImportJobResponse Text

The unique identifier for the application that's associated with the import job.