Network.AWS.Pinpoint.Types.InAppCampaignSchedule

Description

Documentation

data InAppCampaignSchedule

Schedule of the campaign.

See: newInAppCampaignSchedule smart constructor.

Constructors

InAppCampaignSchedule'

Fields

endDate :: Maybe Text

The scheduled time after which the in-app message should not be shown. Timestamp is in ISO 8601 format.

quietTime :: Maybe QuietTime

Time during which the in-app message should not be shown to the user.

eventFilter :: Maybe CampaignEventFilter

The event filter the SDK has to use to show the in-app message in the application.

newInAppCampaignSchedule :: InAppCampaignSchedule

Create a value of InAppCampaignSchedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endDate:InAppCampaignSchedule', inAppCampaignSchedule_endDate - The scheduled time after which the in-app message should not be shown. Timestamp is in ISO 8601 format.

$sel:quietTime:InAppCampaignSchedule', inAppCampaignSchedule_quietTime - Time during which the in-app message should not be shown to the user.

$sel:eventFilter:InAppCampaignSchedule', inAppCampaignSchedule_eventFilter - The event filter the SDK has to use to show the in-app message in the application.

inAppCampaignSchedule_endDate :: Lens' InAppCampaignSchedule (Maybe Text)

The scheduled time after which the in-app message should not be shown. Timestamp is in ISO 8601 format.

inAppCampaignSchedule_quietTime :: Lens' InAppCampaignSchedule (Maybe QuietTime)

Time during which the in-app message should not be shown to the user.

inAppCampaignSchedule_eventFilter :: Lens' InAppCampaignSchedule (Maybe CampaignEventFilter)

The event filter the SDK has to use to show the in-app message in the application.