Network.AWS.Pinpoint.Types.InAppMessageCampaign

Description

Documentation

data InAppMessageCampaign

Targeted in-app message campaign.

See: newInAppMessageCampaign smart constructor.

Constructors

InAppMessageCampaign'

Fields

sessionCap :: Maybe Int

Session cap which controls the number of times an in-app message can be shown to the endpoint during an application session.

campaignId :: Maybe Text

Campaign id of the corresponding campaign.

inAppMessage :: Maybe InAppMessage

In-app message content with all fields required for rendering an in-app message.

priority :: Maybe Int

Priority of the in-app message.

treatmentId :: Maybe Text

Treatment id of the campaign.

dailyCap :: Maybe Int

Daily cap which controls the number of times any in-app messages can be shown to the endpoint during a day.

totalCap :: Maybe Int

Total cap which controls the number of times an in-app message can be shown to the endpoint.

schedule :: Maybe InAppCampaignSchedule

Schedule of the campaign.

newInAppMessageCampaign :: InAppMessageCampaign

Create a value of InAppMessageCampaign with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sessionCap:InAppMessageCampaign', inAppMessageCampaign_sessionCap - Session cap which controls the number of times an in-app message can be shown to the endpoint during an application session.

$sel:campaignId:InAppMessageCampaign', inAppMessageCampaign_campaignId - Campaign id of the corresponding campaign.

$sel:inAppMessage:InAppMessageCampaign', inAppMessageCampaign_inAppMessage - In-app message content with all fields required for rendering an in-app message.

$sel:priority:InAppMessageCampaign', inAppMessageCampaign_priority - Priority of the in-app message.

$sel:treatmentId:InAppMessageCampaign', inAppMessageCampaign_treatmentId - Treatment id of the campaign.

$sel:dailyCap:InAppMessageCampaign', inAppMessageCampaign_dailyCap - Daily cap which controls the number of times any in-app messages can be shown to the endpoint during a day.

$sel:totalCap:InAppMessageCampaign', inAppMessageCampaign_totalCap - Total cap which controls the number of times an in-app message can be shown to the endpoint.

$sel:schedule:InAppMessageCampaign', inAppMessageCampaign_schedule - Schedule of the campaign.

inAppMessageCampaign_sessionCap :: Lens' InAppMessageCampaign (Maybe Int)

Session cap which controls the number of times an in-app message can be shown to the endpoint during an application session.

inAppMessageCampaign_campaignId :: Lens' InAppMessageCampaign (Maybe Text)

Campaign id of the corresponding campaign.

inAppMessageCampaign_inAppMessage :: Lens' InAppMessageCampaign (Maybe InAppMessage)

In-app message content with all fields required for rendering an in-app message.

inAppMessageCampaign_priority :: Lens' InAppMessageCampaign (Maybe Int)

Priority of the in-app message.

inAppMessageCampaign_treatmentId :: Lens' InAppMessageCampaign (Maybe Text)

Treatment id of the campaign.

inAppMessageCampaign_dailyCap :: Lens' InAppMessageCampaign (Maybe Int)

Daily cap which controls the number of times any in-app messages can be shown to the endpoint during a day.

inAppMessageCampaign_totalCap :: Lens' InAppMessageCampaign (Maybe Int)

Total cap which controls the number of times an in-app message can be shown to the endpoint.

inAppMessageCampaign_schedule :: Lens' InAppMessageCampaign (Maybe InAppCampaignSchedule)

Schedule of the campaign.