Network.AWS.Pinpoint.Types.JourneyDateRangeKpiResponse

Description

Documentation

data JourneyDateRangeKpiResponse

Provides the results of a query that retrieved the data for a standard engagement metric that applies to a journey, and provides information about that query.

See: newJourneyDateRangeKpiResponse smart constructor.

Constructors

JourneyDateRangeKpiResponse'

Fields

nextToken :: Maybe Text

The string to use in a subsequent request to get the next page of results in a paginated response. This value is null for the Journey Engagement Metrics resource because the resource returns all results in a single page.

kpiResult :: BaseKpiResult

An array of objects that contains the results of the query. Each object contains the value for the metric and metadata about that value.

kpiName :: Text

The name of the metric, also referred to as a /key performance indicator (KPI)/, that the data was retrieved for. This value describes the associated metric and consists of two or more terms, which are comprised of lowercase alphanumeric characters, separated by a hyphen. For a list of possible values, see the Amazon Pinpoint Developer Guide.

journeyId :: Text

The unique identifier for the journey that the metric applies to.

endTime :: POSIX

The last date and time of the date range that was used to filter the query results, in extended ISO 8601 format. The date range is inclusive.

startTime :: POSIX

The first date and time of the date range that was used to filter the query results, in extended ISO 8601 format. The date range is inclusive.

applicationId :: Text

The unique identifier for the application that the metric applies to.

newJourneyDateRangeKpiResponse

Arguments

:: BaseKpiResult
$sel:kpiResult:JourneyDateRangeKpiResponse'

-> Text
$sel:kpiName:JourneyDateRangeKpiResponse'

-> Text
$sel:journeyId:JourneyDateRangeKpiResponse'

-> UTCTime
$sel:endTime:JourneyDateRangeKpiResponse'

-> UTCTime
$sel:startTime:JourneyDateRangeKpiResponse'

-> Text
$sel:applicationId:JourneyDateRangeKpiResponse'

-> JourneyDateRangeKpiResponse

Create a value of JourneyDateRangeKpiResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:JourneyDateRangeKpiResponse', journeyDateRangeKpiResponse_nextToken - The string to use in a subsequent request to get the next page of results in a paginated response. This value is null for the Journey Engagement Metrics resource because the resource returns all results in a single page.

$sel:kpiResult:JourneyDateRangeKpiResponse', journeyDateRangeKpiResponse_kpiResult - An array of objects that contains the results of the query. Each object contains the value for the metric and metadata about that value.

$sel:kpiName:JourneyDateRangeKpiResponse', journeyDateRangeKpiResponse_kpiName - The name of the metric, also referred to as a /key performance indicator (KPI)/, that the data was retrieved for. This value describes the associated metric and consists of two or more terms, which are comprised of lowercase alphanumeric characters, separated by a hyphen. For a list of possible values, see the Amazon Pinpoint Developer Guide.

$sel:journeyId:JourneyDateRangeKpiResponse', journeyDateRangeKpiResponse_journeyId - The unique identifier for the journey that the metric applies to.

$sel:endTime:JourneyDateRangeKpiResponse', journeyDateRangeKpiResponse_endTime - The last date and time of the date range that was used to filter the query results, in extended ISO 8601 format. The date range is inclusive.

$sel:startTime:JourneyDateRangeKpiResponse', journeyDateRangeKpiResponse_startTime - The first date and time of the date range that was used to filter the query results, in extended ISO 8601 format. The date range is inclusive.

$sel:applicationId:JourneyDateRangeKpiResponse', journeyDateRangeKpiResponse_applicationId - The unique identifier for the application that the metric applies to.

journeyDateRangeKpiResponse_nextToken :: Lens' JourneyDateRangeKpiResponse (Maybe Text)

The string to use in a subsequent request to get the next page of results in a paginated response. This value is null for the Journey Engagement Metrics resource because the resource returns all results in a single page.

journeyDateRangeKpiResponse_kpiResult :: Lens' JourneyDateRangeKpiResponse BaseKpiResult

An array of objects that contains the results of the query. Each object contains the value for the metric and metadata about that value.

journeyDateRangeKpiResponse_kpiName :: Lens' JourneyDateRangeKpiResponse Text

The name of the metric, also referred to as a /key performance indicator (KPI)/, that the data was retrieved for. This value describes the associated metric and consists of two or more terms, which are comprised of lowercase alphanumeric characters, separated by a hyphen. For a list of possible values, see the Amazon Pinpoint Developer Guide.

journeyDateRangeKpiResponse_journeyId :: Lens' JourneyDateRangeKpiResponse Text

The unique identifier for the journey that the metric applies to.

journeyDateRangeKpiResponse_endTime :: Lens' JourneyDateRangeKpiResponse UTCTime

The last date and time of the date range that was used to filter the query results, in extended ISO 8601 format. The date range is inclusive.

journeyDateRangeKpiResponse_startTime :: Lens' JourneyDateRangeKpiResponse UTCTime

The first date and time of the date range that was used to filter the query results, in extended ISO 8601 format. The date range is inclusive.

journeyDateRangeKpiResponse_applicationId :: Lens' JourneyDateRangeKpiResponse Text

The unique identifier for the application that the metric applies to.