Network.AWS.Pinpoint.Types.JourneyExecutionActivityMetricsResponse

Description

Documentation

data JourneyExecutionActivityMetricsResponse

Provides the results of a query that retrieved the data for a standard execution metric that applies to a journey activity, and provides information about that query.

See: newJourneyExecutionActivityMetricsResponse smart constructor.

Constructors

JourneyExecutionActivityMetricsResponse'

Fields

metrics :: HashMap Text Text

A JSON object that contains the results of the query. The results vary depending on the type of activity (ActivityType). For information about the structure and contents of the results, see the Amazon Pinpoint Developer Guide.

journeyId :: Text

The unique identifier for the journey that the metric applies to.

lastEvaluatedTime :: Text

The date and time, in ISO 8601 format, when Amazon Pinpoint last evaluated the execution status of the activity and updated the data for the metric.

journeyActivityId :: Text

The unique identifier for the activity that the metric applies to.

activityType :: Text

The type of activity that the metric applies to. Possible values are:

CONDITIONAL_SPLIT - For a yes/no split activity, which is an activity that sends participants down one of two paths in a journey.

HOLDOUT - For a holdout activity, which is an activity that stops a journey for a specified percentage of participants.

MESSAGE - For an email activity, which is an activity that sends an email message to participants.

MULTI_CONDITIONAL_SPLIT - For a multivariate split activity, which is an activity that sends participants down one of as many as five paths in a journey.

RANDOM_SPLIT - For a random split activity, which is an activity that sends specified percentages of participants down one of as many as five paths in a journey.

WAIT - For a wait activity, which is an activity that waits for a certain amount of time or until a specific date and time before moving participants to the next activity in a journey.

applicationId :: Text

The unique identifier for the application that the metric applies to.

newJourneyExecutionActivityMetricsResponse

Arguments

:: Text
$sel:journeyId:JourneyExecutionActivityMetricsResponse'

-> Text
$sel:lastEvaluatedTime:JourneyExecutionActivityMetricsResponse'

-> Text
$sel:journeyActivityId:JourneyExecutionActivityMetricsResponse'

-> Text
$sel:activityType:JourneyExecutionActivityMetricsResponse'

-> Text
$sel:applicationId:JourneyExecutionActivityMetricsResponse'

-> JourneyExecutionActivityMetricsResponse

Create a value of JourneyExecutionActivityMetricsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:metrics:JourneyExecutionActivityMetricsResponse', journeyExecutionActivityMetricsResponse_metrics - A JSON object that contains the results of the query. The results vary depending on the type of activity (ActivityType). For information about the structure and contents of the results, see the Amazon Pinpoint Developer Guide.

$sel:journeyId:JourneyExecutionActivityMetricsResponse', journeyExecutionActivityMetricsResponse_journeyId - The unique identifier for the journey that the metric applies to.

$sel:lastEvaluatedTime:JourneyExecutionActivityMetricsResponse', journeyExecutionActivityMetricsResponse_lastEvaluatedTime - The date and time, in ISO 8601 format, when Amazon Pinpoint last evaluated the execution status of the activity and updated the data for the metric.

$sel:journeyActivityId:JourneyExecutionActivityMetricsResponse', journeyExecutionActivityMetricsResponse_journeyActivityId - The unique identifier for the activity that the metric applies to.

$sel:activityType:JourneyExecutionActivityMetricsResponse', journeyExecutionActivityMetricsResponse_activityType - The type of activity that the metric applies to. Possible values are:

CONDITIONAL_SPLIT - For a yes/no split activity, which is an activity that sends participants down one of two paths in a journey.

HOLDOUT - For a holdout activity, which is an activity that stops a journey for a specified percentage of participants.

MESSAGE - For an email activity, which is an activity that sends an email message to participants.

MULTI_CONDITIONAL_SPLIT - For a multivariate split activity, which is an activity that sends participants down one of as many as five paths in a journey.

RANDOM_SPLIT - For a random split activity, which is an activity that sends specified percentages of participants down one of as many as five paths in a journey.

WAIT - For a wait activity, which is an activity that waits for a certain amount of time or until a specific date and time before moving participants to the next activity in a journey.

$sel:applicationId:JourneyExecutionActivityMetricsResponse', journeyExecutionActivityMetricsResponse_applicationId - The unique identifier for the application that the metric applies to.

journeyExecutionActivityMetricsResponse_metrics :: Lens' JourneyExecutionActivityMetricsResponse (HashMap Text Text)

A JSON object that contains the results of the query. The results vary depending on the type of activity (ActivityType). For information about the structure and contents of the results, see the Amazon Pinpoint Developer Guide.

journeyExecutionActivityMetricsResponse_journeyId :: Lens' JourneyExecutionActivityMetricsResponse Text

The unique identifier for the journey that the metric applies to.

journeyExecutionActivityMetricsResponse_lastEvaluatedTime :: Lens' JourneyExecutionActivityMetricsResponse Text

The date and time, in ISO 8601 format, when Amazon Pinpoint last evaluated the execution status of the activity and updated the data for the metric.

journeyExecutionActivityMetricsResponse_journeyActivityId :: Lens' JourneyExecutionActivityMetricsResponse Text

The unique identifier for the activity that the metric applies to.

journeyExecutionActivityMetricsResponse_activityType :: Lens' JourneyExecutionActivityMetricsResponse Text

The type of activity that the metric applies to. Possible values are:

CONDITIONAL_SPLIT - For a yes/no split activity, which is an activity that sends participants down one of two paths in a journey.

HOLDOUT - For a holdout activity, which is an activity that stops a journey for a specified percentage of participants.

MESSAGE - For an email activity, which is an activity that sends an email message to participants.

MULTI_CONDITIONAL_SPLIT - For a multivariate split activity, which is an activity that sends participants down one of as many as five paths in a journey.

RANDOM_SPLIT - For a random split activity, which is an activity that sends specified percentages of participants down one of as many as five paths in a journey.

WAIT - For a wait activity, which is an activity that waits for a certain amount of time or until a specific date and time before moving participants to the next activity in a journey.

journeyExecutionActivityMetricsResponse_applicationId :: Lens' JourneyExecutionActivityMetricsResponse Text

The unique identifier for the application that the metric applies to.