Network.AWS.Pinpoint.Types.JourneyLimits

Description

Documentation

data JourneyLimits

Specifies limits on the messages that a journey can send and the number of times participants can enter a journey.

See: newJourneyLimits smart constructor.

Constructors

JourneyLimits'

Fields

messagesPerSecond :: Maybe Int

The maximum number of messages that the journey can send each second.

endpointReentryCap :: Maybe Int

The maximum number of times that a participant can enter the journey. The maximum value is 100. To allow participants to enter the journey an unlimited number of times, set this value to 0.

endpointReentryInterval :: Maybe Text

Minimum time that must pass before an endpoint can re-enter a given journey. The duration should use an ISO 8601 format, such as PT1H.

dailyCap :: Maybe Int

The maximum number of messages that the journey can send to a single participant during a 24-hour period. The maximum value is 100.

newJourneyLimits :: JourneyLimits

Create a value of JourneyLimits with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:messagesPerSecond:JourneyLimits', journeyLimits_messagesPerSecond - The maximum number of messages that the journey can send each second.

$sel:endpointReentryCap:JourneyLimits', journeyLimits_endpointReentryCap - The maximum number of times that a participant can enter the journey. The maximum value is 100. To allow participants to enter the journey an unlimited number of times, set this value to 0.

$sel:endpointReentryInterval:JourneyLimits', journeyLimits_endpointReentryInterval - Minimum time that must pass before an endpoint can re-enter a given journey. The duration should use an ISO 8601 format, such as PT1H.

$sel:dailyCap:JourneyLimits', journeyLimits_dailyCap - The maximum number of messages that the journey can send to a single participant during a 24-hour period. The maximum value is 100.

journeyLimits_messagesPerSecond :: Lens' JourneyLimits (Maybe Int)

The maximum number of messages that the journey can send each second.

journeyLimits_endpointReentryCap :: Lens' JourneyLimits (Maybe Int)

The maximum number of times that a participant can enter the journey. The maximum value is 100. To allow participants to enter the journey an unlimited number of times, set this value to 0.

journeyLimits_endpointReentryInterval :: Lens' JourneyLimits (Maybe Text)

Minimum time that must pass before an endpoint can re-enter a given journey. The duration should use an ISO 8601 format, such as PT1H.

journeyLimits_dailyCap :: Lens' JourneyLimits (Maybe Int)

The maximum number of messages that the journey can send to a single participant during a 24-hour period. The maximum value is 100.