Network.AWS.Pinpoint.Types.JourneySMSMessage

Description

Documentation

data JourneySMSMessage

Specifies the sender ID and message type for an SMS message that's sent to participants in a journey.

See: newJourneySMSMessage smart constructor.

Constructors

JourneySMSMessage'

Fields

entityId :: Maybe Text

The entity ID or Principal Entity (PE) id received from the regulatory body for sending SMS in your country.

senderId :: Maybe Text

The sender ID to display as the sender of the message on a recipient's device. Support for sender IDs varies by country or region. For more information, see Supported Countries and Regions in the Amazon Pinpoint User Guide.

messageType :: Maybe MessageType

The SMS message type. Valid values are TRANSACTIONAL (for messages that are critical or time-sensitive, such as a one-time passwords) and PROMOTIONAL (for messsages that aren't critical or time-sensitive, such as marketing messages).

templateId :: Maybe Text

The template ID received from the regulatory body for sending SMS in your country.

originationNumber :: Maybe Text

The long code to send the SMS message from. This value should be one of the dedicated long codes that's assigned to your AWS account. Although it isn't required, we recommend that you specify the long code using an E.164 format to ensure prompt and accurate delivery of the message. For example, +12065550100.

newJourneySMSMessage :: JourneySMSMessage

Create a value of JourneySMSMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:entityId:JourneySMSMessage', journeySMSMessage_entityId - The entity ID or Principal Entity (PE) id received from the regulatory body for sending SMS in your country.

$sel:senderId:JourneySMSMessage', journeySMSMessage_senderId - The sender ID to display as the sender of the message on a recipient's device. Support for sender IDs varies by country or region. For more information, see Supported Countries and Regions in the Amazon Pinpoint User Guide.

$sel:messageType:JourneySMSMessage', journeySMSMessage_messageType - The SMS message type. Valid values are TRANSACTIONAL (for messages that are critical or time-sensitive, such as a one-time passwords) and PROMOTIONAL (for messsages that aren't critical or time-sensitive, such as marketing messages).

$sel:templateId:JourneySMSMessage', journeySMSMessage_templateId - The template ID received from the regulatory body for sending SMS in your country.

$sel:originationNumber:JourneySMSMessage', journeySMSMessage_originationNumber - The long code to send the SMS message from. This value should be one of the dedicated long codes that's assigned to your AWS account. Although it isn't required, we recommend that you specify the long code using an E.164 format to ensure prompt and accurate delivery of the message. For example, +12065550100.

journeySMSMessage_entityId :: Lens' JourneySMSMessage (Maybe Text)

The entity ID or Principal Entity (PE) id received from the regulatory body for sending SMS in your country.

journeySMSMessage_senderId :: Lens' JourneySMSMessage (Maybe Text)

The sender ID to display as the sender of the message on a recipient's device. Support for sender IDs varies by country or region. For more information, see Supported Countries and Regions in the Amazon Pinpoint User Guide.

journeySMSMessage_messageType :: Lens' JourneySMSMessage (Maybe MessageType)

The SMS message type. Valid values are TRANSACTIONAL (for messages that are critical or time-sensitive, such as a one-time passwords) and PROMOTIONAL (for messsages that aren't critical or time-sensitive, such as marketing messages).

journeySMSMessage_templateId :: Lens' JourneySMSMessage (Maybe Text)

The template ID received from the regulatory body for sending SMS in your country.

journeySMSMessage_originationNumber :: Lens' JourneySMSMessage (Maybe Text)

The long code to send the SMS message from. This value should be one of the dedicated long codes that's assigned to your AWS account. Although it isn't required, we recommend that you specify the long code using an E.164 format to ensure prompt and accurate delivery of the message. For example, +12065550100.