Network.AWS.Pinpoint.Types.JourneysResponse

Description

Documentation

data JourneysResponse

Provides information about the status, configuration, and other settings for all the journeys that are associated with an application.

See: newJourneysResponse smart constructor.

Constructors

JourneysResponse'

Fields

nextToken :: Maybe Text

The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

item :: [JourneyResponse]

An array of responses, one for each journey that's associated with the application.

newJourneysResponse :: JourneysResponse

Create a value of JourneysResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:JourneysResponse', journeysResponse_nextToken - The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

$sel:item:JourneysResponse', journeysResponse_item - An array of responses, one for each journey that's associated with the application.

journeysResponse_nextToken :: Lens' JourneysResponse (Maybe Text)

The string to use in a subsequent request to get the next page of results in a paginated response. This value is null if there are no additional pages.

journeysResponse_item :: Lens' JourneysResponse [JourneyResponse]

An array of responses, one for each journey that's associated with the application.