Network.AWS.Pinpoint.Types.MessageConfiguration

Description

Documentation

data MessageConfiguration

Specifies the message configuration settings for a campaign.

See: newMessageConfiguration smart constructor.

Constructors

MessageConfiguration'

Fields

aDMMessage :: Maybe Message

The message that the campaign sends through the ADM (Amazon Device Messaging) channel. If specified, this message overrides the default message.

defaultMessage :: Maybe Message

The default message that the campaign sends through all the channels that are configured for the campaign.

inAppMessage :: Maybe CampaignInAppMessage

The in-app message configuration.

gCMMessage :: Maybe Message

The message that the campaign sends through the GCM channel, which enables Amazon Pinpoint to send push notifications through the Firebase Cloud Messaging (FCM), formerly Google Cloud Messaging (GCM), service. If specified, this message overrides the default message.

emailMessage :: Maybe CampaignEmailMessage

The message that the campaign sends through the email channel. If specified, this message overrides the default message.

aPNSMessage :: Maybe Message

The message that the campaign sends through the APNs (Apple Push Notification service) channel. If specified, this message overrides the default message.

sMSMessage :: Maybe CampaignSmsMessage

The message that the campaign sends through the SMS channel. If specified, this message overrides the default message.

customMessage :: Maybe CampaignCustomMessage

The message that the campaign sends through a custom channel, as specified by the delivery configuration (CustomDeliveryConfiguration) settings for the campaign. If specified, this message overrides the default message.

baiduMessage :: Maybe Message

The message that the campaign sends through the Baidu (Baidu Cloud Push) channel. If specified, this message overrides the default message.

newMessageConfiguration :: MessageConfiguration

Create a value of MessageConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aDMMessage:MessageConfiguration', messageConfiguration_aDMMessage - The message that the campaign sends through the ADM (Amazon Device Messaging) channel. If specified, this message overrides the default message.

$sel:defaultMessage:MessageConfiguration', messageConfiguration_defaultMessage - The default message that the campaign sends through all the channels that are configured for the campaign.

$sel:inAppMessage:MessageConfiguration', messageConfiguration_inAppMessage - The in-app message configuration.

$sel:gCMMessage:MessageConfiguration', messageConfiguration_gCMMessage - The message that the campaign sends through the GCM channel, which enables Amazon Pinpoint to send push notifications through the Firebase Cloud Messaging (FCM), formerly Google Cloud Messaging (GCM), service. If specified, this message overrides the default message.

$sel:emailMessage:MessageConfiguration', messageConfiguration_emailMessage - The message that the campaign sends through the email channel. If specified, this message overrides the default message.

$sel:aPNSMessage:MessageConfiguration', messageConfiguration_aPNSMessage - The message that the campaign sends through the APNs (Apple Push Notification service) channel. If specified, this message overrides the default message.

$sel:sMSMessage:MessageConfiguration', messageConfiguration_sMSMessage - The message that the campaign sends through the SMS channel. If specified, this message overrides the default message.

$sel:customMessage:MessageConfiguration', messageConfiguration_customMessage - The message that the campaign sends through a custom channel, as specified by the delivery configuration (CustomDeliveryConfiguration) settings for the campaign. If specified, this message overrides the default message.

$sel:baiduMessage:MessageConfiguration', messageConfiguration_baiduMessage - The message that the campaign sends through the Baidu (Baidu Cloud Push) channel. If specified, this message overrides the default message.

messageConfiguration_aDMMessage :: Lens' MessageConfiguration (Maybe Message)

The message that the campaign sends through the ADM (Amazon Device Messaging) channel. If specified, this message overrides the default message.

messageConfiguration_defaultMessage :: Lens' MessageConfiguration (Maybe Message)

The default message that the campaign sends through all the channels that are configured for the campaign.

messageConfiguration_inAppMessage :: Lens' MessageConfiguration (Maybe CampaignInAppMessage)

The in-app message configuration.

messageConfiguration_gCMMessage :: Lens' MessageConfiguration (Maybe Message)

The message that the campaign sends through the GCM channel, which enables Amazon Pinpoint to send push notifications through the Firebase Cloud Messaging (FCM), formerly Google Cloud Messaging (GCM), service. If specified, this message overrides the default message.

messageConfiguration_emailMessage :: Lens' MessageConfiguration (Maybe CampaignEmailMessage)

The message that the campaign sends through the email channel. If specified, this message overrides the default message.

messageConfiguration_aPNSMessage :: Lens' MessageConfiguration (Maybe Message)

The message that the campaign sends through the APNs (Apple Push Notification service) channel. If specified, this message overrides the default message.

messageConfiguration_sMSMessage :: Lens' MessageConfiguration (Maybe CampaignSmsMessage)

The message that the campaign sends through the SMS channel. If specified, this message overrides the default message.

messageConfiguration_customMessage :: Lens' MessageConfiguration (Maybe CampaignCustomMessage)

The message that the campaign sends through a custom channel, as specified by the delivery configuration (CustomDeliveryConfiguration) settings for the campaign. If specified, this message overrides the default message.

messageConfiguration_baiduMessage :: Lens' MessageConfiguration (Maybe Message)

The message that the campaign sends through the Baidu (Baidu Cloud Push) channel. If specified, this message overrides the default message.