Network.AWS.Pinpoint.Types.MessageRequest

Description

Documentation

data MessageRequest

Specifies the configuration and other settings for a message.

See: newMessageRequest smart constructor.

Constructors

MessageRequest'

Fields

endpoints :: Maybe (HashMap Text EndpointSendConfiguration)

A map of key-value pairs, where each key is an endpoint ID and each value is an EndpointSendConfiguration object. You can use an EndpointSendConfiguration object to tailor the message for an endpoint by specifying settings such as content overrides and message variables.

context :: Maybe (HashMap Text Text)

A map of custom attributes to attach to the message. For a push notification, this payload is added to the data.pinpoint object. For an email or text message, this payload is added to email/SMS delivery receipt event attributes.

traceId :: Maybe Text

The unique identifier for tracing the message. This identifier is visible to message recipients.

templateConfiguration :: Maybe TemplateConfiguration

The message template to use for the message.

addresses :: Maybe (HashMap Text AddressConfiguration)

A map of key-value pairs, where each key is an address and each value is an AddressConfiguration object. An address can be a push notification token, a phone number, or an email address. You can use an AddressConfiguration object to tailor the message for an address by specifying settings such as content overrides and message variables.

messageConfiguration :: DirectMessageConfiguration

The settings and content for the default message and any default messages that you defined for specific channels.

newMessageRequest

Arguments

:: DirectMessageConfiguration
$sel:messageConfiguration:MessageRequest'

-> MessageRequest

Create a value of MessageRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpoints:MessageRequest', messageRequest_endpoints - A map of key-value pairs, where each key is an endpoint ID and each value is an EndpointSendConfiguration object. You can use an EndpointSendConfiguration object to tailor the message for an endpoint by specifying settings such as content overrides and message variables.

$sel:context:MessageRequest', messageRequest_context - A map of custom attributes to attach to the message. For a push notification, this payload is added to the data.pinpoint object. For an email or text message, this payload is added to email/SMS delivery receipt event attributes.

$sel:traceId:MessageRequest', messageRequest_traceId - The unique identifier for tracing the message. This identifier is visible to message recipients.

$sel:templateConfiguration:MessageRequest', messageRequest_templateConfiguration - The message template to use for the message.

$sel:addresses:MessageRequest', messageRequest_addresses - A map of key-value pairs, where each key is an address and each value is an AddressConfiguration object. An address can be a push notification token, a phone number, or an email address. You can use an AddressConfiguration object to tailor the message for an address by specifying settings such as content overrides and message variables.

$sel:messageConfiguration:MessageRequest', messageRequest_messageConfiguration - The settings and content for the default message and any default messages that you defined for specific channels.

messageRequest_endpoints :: Lens' MessageRequest (Maybe (HashMap Text EndpointSendConfiguration))

A map of key-value pairs, where each key is an endpoint ID and each value is an EndpointSendConfiguration object. You can use an EndpointSendConfiguration object to tailor the message for an endpoint by specifying settings such as content overrides and message variables.

messageRequest_context :: Lens' MessageRequest (Maybe (HashMap Text Text))

A map of custom attributes to attach to the message. For a push notification, this payload is added to the data.pinpoint object. For an email or text message, this payload is added to email/SMS delivery receipt event attributes.

messageRequest_traceId :: Lens' MessageRequest (Maybe Text)

The unique identifier for tracing the message. This identifier is visible to message recipients.

messageRequest_templateConfiguration :: Lens' MessageRequest (Maybe TemplateConfiguration)

The message template to use for the message.

messageRequest_addresses :: Lens' MessageRequest (Maybe (HashMap Text AddressConfiguration))

A map of key-value pairs, where each key is an address and each value is an AddressConfiguration object. An address can be a push notification token, a phone number, or an email address. You can use an AddressConfiguration object to tailor the message for an address by specifying settings such as content overrides and message variables.

messageRequest_messageConfiguration :: Lens' MessageRequest DirectMessageConfiguration

The settings and content for the default message and any default messages that you defined for specific channels.