Network.AWS.Pinpoint.Types.MultiConditionalBranch

Description

Documentation

data MultiConditionalBranch

Specifies a condition to evaluate for an activity path in a journey.

See: newMultiConditionalBranch smart constructor.

Constructors

MultiConditionalBranch'

Fields

condition :: Maybe SimpleCondition

The condition to evaluate for the activity path.

nextActivity :: Maybe Text

The unique identifier for the next activity to perform, after completing the activity for the path.

newMultiConditionalBranch :: MultiConditionalBranch

Create a value of MultiConditionalBranch with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:condition:MultiConditionalBranch', multiConditionalBranch_condition - The condition to evaluate for the activity path.

$sel:nextActivity:MultiConditionalBranch', multiConditionalBranch_nextActivity - The unique identifier for the next activity to perform, after completing the activity for the path.

multiConditionalBranch_condition :: Lens' MultiConditionalBranch (Maybe SimpleCondition)

The condition to evaluate for the activity path.

multiConditionalBranch_nextActivity :: Lens' MultiConditionalBranch (Maybe Text)

The unique identifier for the next activity to perform, after completing the activity for the path.