Network.AWS.Pinpoint.Types.MultiConditionalSplitActivity

Description

Documentation

data MultiConditionalSplitActivity

Specifies the settings for a multivariate split activity in a journey. This type of activity sends participants down one of as many as five paths (including a default Else path) in a journey, based on conditions that you specify.

To create multivariate split activities that send participants down different paths based on push notification events (such as Open or Received events), your mobile app has to specify the User ID and Endpoint ID values. For more information, see Integrating Amazon Pinpoint with your application in the Amazon Pinpoint Developer Guide.

See: newMultiConditionalSplitActivity smart constructor.

Constructors

MultiConditionalSplitActivity'

Fields

defaultActivity :: Maybe Text

The unique identifier for the activity to perform for participants who don't meet any of the conditions specified for other paths in the activity.

evaluationWaitTime :: Maybe WaitTime

The amount of time to wait or the date and time when Amazon Pinpoint determines whether the conditions are met.

branches :: Maybe [MultiConditionalBranch]

The paths for the activity, including the conditions for entering each path and the activity to perform for each path.

newMultiConditionalSplitActivity :: MultiConditionalSplitActivity

Create a value of MultiConditionalSplitActivity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:defaultActivity:MultiConditionalSplitActivity', multiConditionalSplitActivity_defaultActivity - The unique identifier for the activity to perform for participants who don't meet any of the conditions specified for other paths in the activity.

$sel:evaluationWaitTime:MultiConditionalSplitActivity', multiConditionalSplitActivity_evaluationWaitTime - The amount of time to wait or the date and time when Amazon Pinpoint determines whether the conditions are met.

$sel:branches:MultiConditionalSplitActivity', multiConditionalSplitActivity_branches - The paths for the activity, including the conditions for entering each path and the activity to perform for each path.

multiConditionalSplitActivity_defaultActivity :: Lens' MultiConditionalSplitActivity (Maybe Text)

The unique identifier for the activity to perform for participants who don't meet any of the conditions specified for other paths in the activity.

multiConditionalSplitActivity_evaluationWaitTime :: Lens' MultiConditionalSplitActivity (Maybe WaitTime)

The amount of time to wait or the date and time when Amazon Pinpoint determines whether the conditions are met.

multiConditionalSplitActivity_branches :: Lens' MultiConditionalSplitActivity (Maybe [MultiConditionalBranch])

The paths for the activity, including the conditions for entering each path and the activity to perform for each path.