Network.AWS.Pinpoint.Types.NumberValidateResponse

Description

Documentation

data NumberValidateResponse

Provides information about a phone number.

See: newNumberValidateResponse smart constructor.

Constructors

NumberValidateResponse'

Fields

phoneType :: Maybe Text

The description of the phone type. Valid values are: MOBILE, LANDLINE, VOIP, INVALID, PREPAID, and OTHER.

zipCode :: Maybe Text

The postal or ZIP code for the location where the phone number was originally registered.

originalPhoneNumber :: Maybe Text

The phone number that was sent in the request body.

originalCountryCodeIso2 :: Maybe Text

The two-character code, in ISO 3166-1 alpha-2 format, that was sent in the request body.

countryCodeIso2 :: Maybe Text

The two-character code, in ISO 3166-1 alpha-2 format, for the country or region where the phone number was originally registered.

county :: Maybe Text

The name of the county where the phone number was originally registered.

city :: Maybe Text

The name of the city where the phone number was originally registered.

carrier :: Maybe Text

The carrier or service provider that the phone number is currently registered with. In some countries and regions, this value may be the carrier or service provider that the phone number was originally registered with.

phoneTypeCode :: Maybe Int

The phone type, represented by an integer. Valid values are: 0 (mobile), 1 (landline), 2 (VoIP), 3 (invalid), 4 (other), and 5 (prepaid).

cleansedPhoneNumberE164 :: Maybe Text

The cleansed phone number, in E.164 format, for the location where the phone number was originally registered.

countryCodeNumeric :: Maybe Text

The numeric code for the country or region where the phone number was originally registered.

cleansedPhoneNumberNational :: Maybe Text

The cleansed phone number, in the format for the location where the phone number was originally registered.

timezone :: Maybe Text

The time zone for the location where the phone number was originally registered.

country :: Maybe Text

The name of the country or region where the phone number was originally registered.

newNumberValidateResponse :: NumberValidateResponse

Create a value of NumberValidateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:phoneType:NumberValidateResponse', numberValidateResponse_phoneType - The description of the phone type. Valid values are: MOBILE, LANDLINE, VOIP, INVALID, PREPAID, and OTHER.

$sel:zipCode:NumberValidateResponse', numberValidateResponse_zipCode - The postal or ZIP code for the location where the phone number was originally registered.

$sel:originalPhoneNumber:NumberValidateResponse', numberValidateResponse_originalPhoneNumber - The phone number that was sent in the request body.

$sel:originalCountryCodeIso2:NumberValidateResponse', numberValidateResponse_originalCountryCodeIso2 - The two-character code, in ISO 3166-1 alpha-2 format, that was sent in the request body.

$sel:countryCodeIso2:NumberValidateResponse', numberValidateResponse_countryCodeIso2 - The two-character code, in ISO 3166-1 alpha-2 format, for the country or region where the phone number was originally registered.

$sel:county:NumberValidateResponse', numberValidateResponse_county - The name of the county where the phone number was originally registered.

$sel:city:NumberValidateResponse', numberValidateResponse_city - The name of the city where the phone number was originally registered.

$sel:carrier:NumberValidateResponse', numberValidateResponse_carrier - The carrier or service provider that the phone number is currently registered with. In some countries and regions, this value may be the carrier or service provider that the phone number was originally registered with.

$sel:phoneTypeCode:NumberValidateResponse', numberValidateResponse_phoneTypeCode - The phone type, represented by an integer. Valid values are: 0 (mobile), 1 (landline), 2 (VoIP), 3 (invalid), 4 (other), and 5 (prepaid).

$sel:cleansedPhoneNumberE164:NumberValidateResponse', numberValidateResponse_cleansedPhoneNumberE164 - The cleansed phone number, in E.164 format, for the location where the phone number was originally registered.

$sel:countryCodeNumeric:NumberValidateResponse', numberValidateResponse_countryCodeNumeric - The numeric code for the country or region where the phone number was originally registered.

$sel:cleansedPhoneNumberNational:NumberValidateResponse', numberValidateResponse_cleansedPhoneNumberNational - The cleansed phone number, in the format for the location where the phone number was originally registered.

$sel:timezone:NumberValidateResponse', numberValidateResponse_timezone - The time zone for the location where the phone number was originally registered.

$sel:country:NumberValidateResponse', numberValidateResponse_country - The name of the country or region where the phone number was originally registered.

numberValidateResponse_phoneType :: Lens' NumberValidateResponse (Maybe Text)

The description of the phone type. Valid values are: MOBILE, LANDLINE, VOIP, INVALID, PREPAID, and OTHER.

numberValidateResponse_zipCode :: Lens' NumberValidateResponse (Maybe Text)

The postal or ZIP code for the location where the phone number was originally registered.

numberValidateResponse_originalPhoneNumber :: Lens' NumberValidateResponse (Maybe Text)

The phone number that was sent in the request body.

numberValidateResponse_originalCountryCodeIso2 :: Lens' NumberValidateResponse (Maybe Text)

The two-character code, in ISO 3166-1 alpha-2 format, that was sent in the request body.

numberValidateResponse_countryCodeIso2 :: Lens' NumberValidateResponse (Maybe Text)

The two-character code, in ISO 3166-1 alpha-2 format, for the country or region where the phone number was originally registered.

numberValidateResponse_county :: Lens' NumberValidateResponse (Maybe Text)

The name of the county where the phone number was originally registered.

numberValidateResponse_city :: Lens' NumberValidateResponse (Maybe Text)

The name of the city where the phone number was originally registered.

numberValidateResponse_carrier :: Lens' NumberValidateResponse (Maybe Text)

The carrier or service provider that the phone number is currently registered with. In some countries and regions, this value may be the carrier or service provider that the phone number was originally registered with.

numberValidateResponse_phoneTypeCode :: Lens' NumberValidateResponse (Maybe Int)

The phone type, represented by an integer. Valid values are: 0 (mobile), 1 (landline), 2 (VoIP), 3 (invalid), 4 (other), and 5 (prepaid).

numberValidateResponse_cleansedPhoneNumberE164 :: Lens' NumberValidateResponse (Maybe Text)

The cleansed phone number, in E.164 format, for the location where the phone number was originally registered.

numberValidateResponse_countryCodeNumeric :: Lens' NumberValidateResponse (Maybe Text)

The numeric code for the country or region where the phone number was originally registered.

numberValidateResponse_cleansedPhoneNumberNational :: Lens' NumberValidateResponse (Maybe Text)

The cleansed phone number, in the format for the location where the phone number was originally registered.

numberValidateResponse_timezone :: Lens' NumberValidateResponse (Maybe Text)

The time zone for the location where the phone number was originally registered.

numberValidateResponse_country :: Lens' NumberValidateResponse (Maybe Text)

The name of the country or region where the phone number was originally registered.