Network.AWS.Pinpoint.Types.PushNotificationTemplateRequest

Description

Documentation

data PushNotificationTemplateRequest

Specifies the content and settings for a message template that can be used in messages that are sent through a push notification channel.

See: newPushNotificationTemplateRequest smart constructor.

Constructors

PushNotificationTemplateRequest'

Fields

templateDescription :: Maybe Text

A custom description of the message template.

baidu :: Maybe AndroidPushNotificationTemplate

The message template to use for the Baidu (Baidu Cloud Push) channel. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

adm :: Maybe AndroidPushNotificationTemplate

The message template to use for the ADM (Amazon Device Messaging) channel. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

defaultSubstitutions :: Maybe Text

A JSON object that specifies the default values to use for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable. When you create a message that's based on the template, you can override these defaults with message-specific and address-specific variables and values.

apns :: Maybe APNSPushNotificationTemplate

The message template to use for the APNs (Apple Push Notification service) channel. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

tags :: Maybe (HashMap Text Text)

A string-to-string map of key-value pairs that defines the tags to associate with the message template. Each tag consists of a required tag key and an associated tag value.

gcm :: Maybe AndroidPushNotificationTemplate

The message template to use for the GCM channel, which is used to send notifications through the Firebase Cloud Messaging (FCM), formerly Google Cloud Messaging (GCM), service. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

recommenderId :: Maybe Text

The unique identifier for the recommender model to use for the message template. Amazon Pinpoint uses this value to determine how to retrieve and process data from a recommender model when it sends messages that use the template, if the template contains message variables for recommendation data.

default' :: Maybe DefaultPushNotificationTemplate

The default message template to use for push notification channels.

newPushNotificationTemplateRequest :: PushNotificationTemplateRequest

Create a value of PushNotificationTemplateRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateDescription:PushNotificationTemplateRequest', pushNotificationTemplateRequest_templateDescription - A custom description of the message template.

$sel:baidu:PushNotificationTemplateRequest', pushNotificationTemplateRequest_baidu - The message template to use for the Baidu (Baidu Cloud Push) channel. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

$sel:adm:PushNotificationTemplateRequest', pushNotificationTemplateRequest_adm - The message template to use for the ADM (Amazon Device Messaging) channel. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

$sel:defaultSubstitutions:PushNotificationTemplateRequest', pushNotificationTemplateRequest_defaultSubstitutions - A JSON object that specifies the default values to use for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable. When you create a message that's based on the template, you can override these defaults with message-specific and address-specific variables and values.

$sel:apns:PushNotificationTemplateRequest', pushNotificationTemplateRequest_apns - The message template to use for the APNs (Apple Push Notification service) channel. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

$sel:tags:PushNotificationTemplateRequest', pushNotificationTemplateRequest_tags - A string-to-string map of key-value pairs that defines the tags to associate with the message template. Each tag consists of a required tag key and an associated tag value.

$sel:gcm:PushNotificationTemplateRequest', pushNotificationTemplateRequest_gcm - The message template to use for the GCM channel, which is used to send notifications through the Firebase Cloud Messaging (FCM), formerly Google Cloud Messaging (GCM), service. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

$sel:recommenderId:PushNotificationTemplateRequest', pushNotificationTemplateRequest_recommenderId - The unique identifier for the recommender model to use for the message template. Amazon Pinpoint uses this value to determine how to retrieve and process data from a recommender model when it sends messages that use the template, if the template contains message variables for recommendation data.

$sel:default':PushNotificationTemplateRequest', pushNotificationTemplateRequest_default - The default message template to use for push notification channels.

pushNotificationTemplateRequest_templateDescription :: Lens' PushNotificationTemplateRequest (Maybe Text)

A custom description of the message template.

pushNotificationTemplateRequest_baidu :: Lens' PushNotificationTemplateRequest (Maybe AndroidPushNotificationTemplate)

The message template to use for the Baidu (Baidu Cloud Push) channel. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

pushNotificationTemplateRequest_adm :: Lens' PushNotificationTemplateRequest (Maybe AndroidPushNotificationTemplate)

The message template to use for the ADM (Amazon Device Messaging) channel. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

pushNotificationTemplateRequest_defaultSubstitutions :: Lens' PushNotificationTemplateRequest (Maybe Text)

A JSON object that specifies the default values to use for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable. When you create a message that's based on the template, you can override these defaults with message-specific and address-specific variables and values.

pushNotificationTemplateRequest_apns :: Lens' PushNotificationTemplateRequest (Maybe APNSPushNotificationTemplate)

The message template to use for the APNs (Apple Push Notification service) channel. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

pushNotificationTemplateRequest_tags :: Lens' PushNotificationTemplateRequest (Maybe (HashMap Text Text))

A string-to-string map of key-value pairs that defines the tags to associate with the message template. Each tag consists of a required tag key and an associated tag value.

pushNotificationTemplateRequest_gcm :: Lens' PushNotificationTemplateRequest (Maybe AndroidPushNotificationTemplate)

The message template to use for the GCM channel, which is used to send notifications through the Firebase Cloud Messaging (FCM), formerly Google Cloud Messaging (GCM), service. This message template overrides the default template for push notification channels (DefaultPushNotificationTemplate).

pushNotificationTemplateRequest_recommenderId :: Lens' PushNotificationTemplateRequest (Maybe Text)

The unique identifier for the recommender model to use for the message template. Amazon Pinpoint uses this value to determine how to retrieve and process data from a recommender model when it sends messages that use the template, if the template contains message variables for recommendation data.

pushNotificationTemplateRequest_default :: Lens' PushNotificationTemplateRequest (Maybe DefaultPushNotificationTemplate)

The default message template to use for push notification channels.