Network.AWS.Pinpoint.Types.RecommenderConfigurationResponse

Description

Documentation

data RecommenderConfigurationResponse

Provides information about Amazon Pinpoint configuration settings for retrieving and processing data from a recommender model.

See: newRecommenderConfigurationResponse smart constructor.

Constructors

RecommenderConfigurationResponse'

Fields

recommendationsDisplayName :: Maybe Text

The custom display name for the standard endpoint or user attribute (RecommendationItems) that temporarily stores recommended items for each endpoint or user, depending on the value for the RecommendationProviderIdType property. This name appears in the Attribute finder of the template editor on the Amazon Pinpoint console.

This value is null if the configuration doesn't invoke an AWS Lambda function (RecommendationTransformerUri) to perform additional processing of recommendation data.

recommendationTransformerUri :: Maybe Text

The name or Amazon Resource Name (ARN) of the AWS Lambda function that Amazon Pinpoint invokes to perform additional processing of recommendation data that it retrieves from the recommender model.

name :: Maybe Text

The custom name of the configuration for the recommender model.

attributes :: Maybe (HashMap Text Text)

A map that defines 1-10 custom endpoint or user attributes, depending on the value for the RecommendationProviderIdType property. Each of these attributes temporarily stores a recommended item that's retrieved from the recommender model and sent to an AWS Lambda function for additional processing. Each attribute can be used as a message variable in a message template.

This value is null if the configuration doesn't invoke an AWS Lambda function (RecommendationTransformerUri) to perform additional processing of recommendation data.

description :: Maybe Text

The custom description of the configuration for the recommender model.

recommendationsPerMessage :: Maybe Int

The number of recommended items that are retrieved from the model for each endpoint or user, depending on the value for the RecommendationProviderIdType property. This number determines how many recommended items are available for use in message variables.

recommendationProviderIdType :: Maybe Text

The type of Amazon Pinpoint ID that's associated with unique user IDs in the recommender model. This value enables the model to use attribute and event data that’s specific to a particular endpoint or user in an Amazon Pinpoint application. Possible values are:

PINPOINT_ENDPOINT_ID - Each user in the model is associated with a particular endpoint in Amazon Pinpoint. The data is correlated based on endpoint IDs in Amazon Pinpoint. This is the default value.

PINPOINT_USER_ID - Each user in the model is associated with a particular user and endpoint in Amazon Pinpoint. The data is correlated based on user IDs in Amazon Pinpoint. If this value is specified, an endpoint definition in Amazon Pinpoint has to specify both a user ID (UserId) and an endpoint ID. Otherwise, messages won’t be sent to the user's endpoint.

recommendationProviderUri :: Text

The Amazon Resource Name (ARN) of the recommender model that Amazon Pinpoint retrieves the recommendation data from. This value is the ARN of an Amazon Personalize campaign.

lastModifiedDate :: Text

The date, in extended ISO 8601 format, when the configuration for the recommender model was last modified.

creationDate :: Text

The date, in extended ISO 8601 format, when the configuration was created for the recommender model.

recommendationProviderRoleArn :: Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that authorizes Amazon Pinpoint to retrieve recommendation data from the recommender model.

id :: Text

The unique identifier for the recommender model configuration.

newRecommenderConfigurationResponse

Arguments

:: Text
$sel:recommendationProviderUri:RecommenderConfigurationResponse'

-> Text
$sel:lastModifiedDate:RecommenderConfigurationResponse'

-> Text
$sel:creationDate:RecommenderConfigurationResponse'

-> Text
$sel:recommendationProviderRoleArn:RecommenderConfigurationResponse'

-> Text
$sel:id:RecommenderConfigurationResponse'

-> RecommenderConfigurationResponse

Create a value of RecommenderConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:recommendationsDisplayName:RecommenderConfigurationResponse', recommenderConfigurationResponse_recommendationsDisplayName - The custom display name for the standard endpoint or user attribute (RecommendationItems) that temporarily stores recommended items for each endpoint or user, depending on the value for the RecommendationProviderIdType property. This name appears in the Attribute finder of the template editor on the Amazon Pinpoint console.

This value is null if the configuration doesn't invoke an AWS Lambda function (RecommendationTransformerUri) to perform additional processing of recommendation data.

$sel:recommendationTransformerUri:RecommenderConfigurationResponse', recommenderConfigurationResponse_recommendationTransformerUri - The name or Amazon Resource Name (ARN) of the AWS Lambda function that Amazon Pinpoint invokes to perform additional processing of recommendation data that it retrieves from the recommender model.

$sel:name:RecommenderConfigurationResponse', recommenderConfigurationResponse_name - The custom name of the configuration for the recommender model.

$sel:attributes:RecommenderConfigurationResponse', recommenderConfigurationResponse_attributes - A map that defines 1-10 custom endpoint or user attributes, depending on the value for the RecommendationProviderIdType property. Each of these attributes temporarily stores a recommended item that's retrieved from the recommender model and sent to an AWS Lambda function for additional processing. Each attribute can be used as a message variable in a message template.

This value is null if the configuration doesn't invoke an AWS Lambda function (RecommendationTransformerUri) to perform additional processing of recommendation data.

$sel:description:RecommenderConfigurationResponse', recommenderConfigurationResponse_description - The custom description of the configuration for the recommender model.

$sel:recommendationsPerMessage:RecommenderConfigurationResponse', recommenderConfigurationResponse_recommendationsPerMessage - The number of recommended items that are retrieved from the model for each endpoint or user, depending on the value for the RecommendationProviderIdType property. This number determines how many recommended items are available for use in message variables.

$sel:recommendationProviderIdType:RecommenderConfigurationResponse', recommenderConfigurationResponse_recommendationProviderIdType - The type of Amazon Pinpoint ID that's associated with unique user IDs in the recommender model. This value enables the model to use attribute and event data that’s specific to a particular endpoint or user in an Amazon Pinpoint application. Possible values are:

PINPOINT_ENDPOINT_ID - Each user in the model is associated with a particular endpoint in Amazon Pinpoint. The data is correlated based on endpoint IDs in Amazon Pinpoint. This is the default value.

PINPOINT_USER_ID - Each user in the model is associated with a particular user and endpoint in Amazon Pinpoint. The data is correlated based on user IDs in Amazon Pinpoint. If this value is specified, an endpoint definition in Amazon Pinpoint has to specify both a user ID (UserId) and an endpoint ID. Otherwise, messages won’t be sent to the user's endpoint.

$sel:recommendationProviderUri:RecommenderConfigurationResponse', recommenderConfigurationResponse_recommendationProviderUri - The Amazon Resource Name (ARN) of the recommender model that Amazon Pinpoint retrieves the recommendation data from. This value is the ARN of an Amazon Personalize campaign.

$sel:lastModifiedDate:RecommenderConfigurationResponse', recommenderConfigurationResponse_lastModifiedDate - The date, in extended ISO 8601 format, when the configuration for the recommender model was last modified.

$sel:creationDate:RecommenderConfigurationResponse', recommenderConfigurationResponse_creationDate - The date, in extended ISO 8601 format, when the configuration was created for the recommender model.

$sel:recommendationProviderRoleArn:RecommenderConfigurationResponse', recommenderConfigurationResponse_recommendationProviderRoleArn - The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that authorizes Amazon Pinpoint to retrieve recommendation data from the recommender model.

$sel:id:RecommenderConfigurationResponse', recommenderConfigurationResponse_id - The unique identifier for the recommender model configuration.

recommenderConfigurationResponse_recommendationsDisplayName :: Lens' RecommenderConfigurationResponse (Maybe Text)

The custom display name for the standard endpoint or user attribute (RecommendationItems) that temporarily stores recommended items for each endpoint or user, depending on the value for the RecommendationProviderIdType property. This name appears in the Attribute finder of the template editor on the Amazon Pinpoint console.

This value is null if the configuration doesn't invoke an AWS Lambda function (RecommendationTransformerUri) to perform additional processing of recommendation data.

recommenderConfigurationResponse_recommendationTransformerUri :: Lens' RecommenderConfigurationResponse (Maybe Text)

The name or Amazon Resource Name (ARN) of the AWS Lambda function that Amazon Pinpoint invokes to perform additional processing of recommendation data that it retrieves from the recommender model.

recommenderConfigurationResponse_name :: Lens' RecommenderConfigurationResponse (Maybe Text)

The custom name of the configuration for the recommender model.

recommenderConfigurationResponse_attributes :: Lens' RecommenderConfigurationResponse (Maybe (HashMap Text Text))

A map that defines 1-10 custom endpoint or user attributes, depending on the value for the RecommendationProviderIdType property. Each of these attributes temporarily stores a recommended item that's retrieved from the recommender model and sent to an AWS Lambda function for additional processing. Each attribute can be used as a message variable in a message template.

This value is null if the configuration doesn't invoke an AWS Lambda function (RecommendationTransformerUri) to perform additional processing of recommendation data.

recommenderConfigurationResponse_description :: Lens' RecommenderConfigurationResponse (Maybe Text)

The custom description of the configuration for the recommender model.

recommenderConfigurationResponse_recommendationsPerMessage :: Lens' RecommenderConfigurationResponse (Maybe Int)

The number of recommended items that are retrieved from the model for each endpoint or user, depending on the value for the RecommendationProviderIdType property. This number determines how many recommended items are available for use in message variables.

recommenderConfigurationResponse_recommendationProviderIdType :: Lens' RecommenderConfigurationResponse (Maybe Text)

The type of Amazon Pinpoint ID that's associated with unique user IDs in the recommender model. This value enables the model to use attribute and event data that’s specific to a particular endpoint or user in an Amazon Pinpoint application. Possible values are:

PINPOINT_ENDPOINT_ID - Each user in the model is associated with a particular endpoint in Amazon Pinpoint. The data is correlated based on endpoint IDs in Amazon Pinpoint. This is the default value.

PINPOINT_USER_ID - Each user in the model is associated with a particular user and endpoint in Amazon Pinpoint. The data is correlated based on user IDs in Amazon Pinpoint. If this value is specified, an endpoint definition in Amazon Pinpoint has to specify both a user ID (UserId) and an endpoint ID. Otherwise, messages won’t be sent to the user's endpoint.

recommenderConfigurationResponse_recommendationProviderUri :: Lens' RecommenderConfigurationResponse Text

The Amazon Resource Name (ARN) of the recommender model that Amazon Pinpoint retrieves the recommendation data from. This value is the ARN of an Amazon Personalize campaign.

recommenderConfigurationResponse_lastModifiedDate :: Lens' RecommenderConfigurationResponse Text

The date, in extended ISO 8601 format, when the configuration for the recommender model was last modified.

recommenderConfigurationResponse_creationDate :: Lens' RecommenderConfigurationResponse Text

The date, in extended ISO 8601 format, when the configuration was created for the recommender model.

recommenderConfigurationResponse_recommendationProviderRoleArn :: Lens' RecommenderConfigurationResponse Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that authorizes Amazon Pinpoint to retrieve recommendation data from the recommender model.

recommenderConfigurationResponse_id :: Lens' RecommenderConfigurationResponse Text

The unique identifier for the recommender model configuration.