Network.AWS.Pinpoint.Types.SMSMessage

Description

Documentation

data SMSMessage

Specifies the default settings for a one-time SMS message that's sent directly to an endpoint.

See: newSMSMessage smart constructor.

Constructors

SMSMessage'

Fields

keyword :: Maybe Text

The SMS program name that you provided to AWS Support when you requested your dedicated number.

entityId :: Maybe Text

The entity ID or Principal Entity (PE) id received from the regulatory body for sending SMS in your country.

body :: Maybe Text

The body of the SMS message.

mediaUrl :: Maybe Text

This field is reserved for future use.

substitutions :: Maybe (HashMap Text [Text])

The message variables to use in the SMS message. You can override the default variables with individual address variables.

senderId :: Maybe Text

The sender ID to display as the sender of the message on a recipient's device. Support for sender IDs varies by country or region.

messageType :: Maybe MessageType

The SMS message type. Valid values are TRANSACTIONAL (for messages that are critical or time-sensitive, such as a one-time passwords) and PROMOTIONAL (for messsages that aren't critical or time-sensitive, such as marketing messages).

templateId :: Maybe Text

The template ID received from the regulatory body for sending SMS in your country.

originationNumber :: Maybe Text

The number to send the SMS message from. This value should be one of the dedicated long or short codes that's assigned to your AWS account. If you don't specify a long or short code, Amazon Pinpoint assigns a random long code to the SMS message and sends the message from that code.

newSMSMessage :: SMSMessage

Create a value of SMSMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:keyword:SMSMessage', sMSMessage_keyword - The SMS program name that you provided to AWS Support when you requested your dedicated number.

$sel:entityId:SMSMessage', sMSMessage_entityId - The entity ID or Principal Entity (PE) id received from the regulatory body for sending SMS in your country.

$sel:body:SMSMessage', sMSMessage_body - The body of the SMS message.

$sel:mediaUrl:SMSMessage', sMSMessage_mediaUrl - This field is reserved for future use.

$sel:substitutions:SMSMessage', sMSMessage_substitutions - The message variables to use in the SMS message. You can override the default variables with individual address variables.

$sel:senderId:SMSMessage', sMSMessage_senderId - The sender ID to display as the sender of the message on a recipient's device. Support for sender IDs varies by country or region.

$sel:messageType:SMSMessage', sMSMessage_messageType - The SMS message type. Valid values are TRANSACTIONAL (for messages that are critical or time-sensitive, such as a one-time passwords) and PROMOTIONAL (for messsages that aren't critical or time-sensitive, such as marketing messages).

$sel:templateId:SMSMessage', sMSMessage_templateId - The template ID received from the regulatory body for sending SMS in your country.

$sel:originationNumber:SMSMessage', sMSMessage_originationNumber - The number to send the SMS message from. This value should be one of the dedicated long or short codes that's assigned to your AWS account. If you don't specify a long or short code, Amazon Pinpoint assigns a random long code to the SMS message and sends the message from that code.

sMSMessage_keyword :: Lens' SMSMessage (Maybe Text)

The SMS program name that you provided to AWS Support when you requested your dedicated number.

sMSMessage_entityId :: Lens' SMSMessage (Maybe Text)

The entity ID or Principal Entity (PE) id received from the regulatory body for sending SMS in your country.

sMSMessage_body :: Lens' SMSMessage (Maybe Text)

The body of the SMS message.

sMSMessage_mediaUrl :: Lens' SMSMessage (Maybe Text)

This field is reserved for future use.

sMSMessage_substitutions :: Lens' SMSMessage (Maybe (HashMap Text [Text]))

The message variables to use in the SMS message. You can override the default variables with individual address variables.

sMSMessage_senderId :: Lens' SMSMessage (Maybe Text)

The sender ID to display as the sender of the message on a recipient's device. Support for sender IDs varies by country or region.

sMSMessage_messageType :: Lens' SMSMessage (Maybe MessageType)

The SMS message type. Valid values are TRANSACTIONAL (for messages that are critical or time-sensitive, such as a one-time passwords) and PROMOTIONAL (for messsages that aren't critical or time-sensitive, such as marketing messages).

sMSMessage_templateId :: Lens' SMSMessage (Maybe Text)

The template ID received from the regulatory body for sending SMS in your country.

sMSMessage_originationNumber :: Lens' SMSMessage (Maybe Text)

The number to send the SMS message from. This value should be one of the dedicated long or short codes that's assigned to your AWS account. If you don't specify a long or short code, Amazon Pinpoint assigns a random long code to the SMS message and sends the message from that code.