Network.AWS.Pinpoint.Types.SMSMessageActivity

Description

Documentation

data SMSMessageActivity

Specifies the settings for an SMS activity in a journey. This type of activity sends a text message to participants.

See: newSMSMessageActivity smart constructor.

Constructors

SMSMessageActivity'

Fields

templateName :: Maybe Text

The name of the SMS message template to use for the message. If specified, this value must match the name of an existing message template.

templateVersion :: Maybe Text

The unique identifier for the version of the SMS template to use for the message. If specified, this value must match the identifier for an existing template version. To retrieve a list of versions and version identifiers for a template, use the Template Versions resource.

If you don't specify a value for this property, Amazon Pinpoint uses the active version of the template. The active version is typically the version of a template that's been most recently reviewed and approved for use, depending on your workflow. It isn't necessarily the latest version of a template.

nextActivity :: Maybe Text

The unique identifier for the next activity to perform, after the message is sent.

messageConfig :: Maybe JourneySMSMessage

Specifies the sender ID and message type for an SMS message that's sent to participants in a journey.

newSMSMessageActivity :: SMSMessageActivity

Create a value of SMSMessageActivity with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateName:SMSMessageActivity', sMSMessageActivity_templateName - The name of the SMS message template to use for the message. If specified, this value must match the name of an existing message template.

$sel:templateVersion:SMSMessageActivity', sMSMessageActivity_templateVersion - The unique identifier for the version of the SMS template to use for the message. If specified, this value must match the identifier for an existing template version. To retrieve a list of versions and version identifiers for a template, use the Template Versions resource.

If you don't specify a value for this property, Amazon Pinpoint uses the active version of the template. The active version is typically the version of a template that's been most recently reviewed and approved for use, depending on your workflow. It isn't necessarily the latest version of a template.

$sel:nextActivity:SMSMessageActivity', sMSMessageActivity_nextActivity - The unique identifier for the next activity to perform, after the message is sent.

$sel:messageConfig:SMSMessageActivity', sMSMessageActivity_messageConfig - Specifies the sender ID and message type for an SMS message that's sent to participants in a journey.

sMSMessageActivity_templateName :: Lens' SMSMessageActivity (Maybe Text)

The name of the SMS message template to use for the message. If specified, this value must match the name of an existing message template.

sMSMessageActivity_templateVersion :: Lens' SMSMessageActivity (Maybe Text)

The unique identifier for the version of the SMS template to use for the message. If specified, this value must match the identifier for an existing template version. To retrieve a list of versions and version identifiers for a template, use the Template Versions resource.

If you don't specify a value for this property, Amazon Pinpoint uses the active version of the template. The active version is typically the version of a template that's been most recently reviewed and approved for use, depending on your workflow. It isn't necessarily the latest version of a template.

sMSMessageActivity_nextActivity :: Lens' SMSMessageActivity (Maybe Text)

The unique identifier for the next activity to perform, after the message is sent.

sMSMessageActivity_messageConfig :: Lens' SMSMessageActivity (Maybe JourneySMSMessage)

Specifies the sender ID and message type for an SMS message that's sent to participants in a journey.