Network.AWS.Pinpoint.Types.SMSTemplateRequest

Description

Documentation

data SMSTemplateRequest

Specifies the content and settings for a message template that can be used in text messages that are sent through the SMS channel.

See: newSMSTemplateRequest smart constructor.

Constructors

SMSTemplateRequest'

Fields

body :: Maybe Text

The message body to use in text messages that are based on the message template.

templateDescription :: Maybe Text

A custom description of the message template.

defaultSubstitutions :: Maybe Text

A JSON object that specifies the default values to use for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable. When you create a message that's based on the template, you can override these defaults with message-specific and address-specific variables and values.

recommenderId :: Maybe Text

The unique identifier for the recommender model to use for the message template. Amazon Pinpoint uses this value to determine how to retrieve and process data from a recommender model when it sends messages that use the template, if the template contains message variables for recommendation data.

tags :: Maybe (HashMap Text Text)

A string-to-string map of key-value pairs that defines the tags to associate with the message template. Each tag consists of a required tag key and an associated tag value.

newSMSTemplateRequest :: SMSTemplateRequest

Create a value of SMSTemplateRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:body:SMSTemplateRequest', sMSTemplateRequest_body - The message body to use in text messages that are based on the message template.

$sel:templateDescription:SMSTemplateRequest', sMSTemplateRequest_templateDescription - A custom description of the message template.

$sel:defaultSubstitutions:SMSTemplateRequest', sMSTemplateRequest_defaultSubstitutions - A JSON object that specifies the default values to use for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable. When you create a message that's based on the template, you can override these defaults with message-specific and address-specific variables and values.

$sel:recommenderId:SMSTemplateRequest', sMSTemplateRequest_recommenderId - The unique identifier for the recommender model to use for the message template. Amazon Pinpoint uses this value to determine how to retrieve and process data from a recommender model when it sends messages that use the template, if the template contains message variables for recommendation data.

$sel:tags:SMSTemplateRequest', sMSTemplateRequest_tags - A string-to-string map of key-value pairs that defines the tags to associate with the message template. Each tag consists of a required tag key and an associated tag value.

sMSTemplateRequest_body :: Lens' SMSTemplateRequest (Maybe Text)

The message body to use in text messages that are based on the message template.

sMSTemplateRequest_templateDescription :: Lens' SMSTemplateRequest (Maybe Text)

A custom description of the message template.

sMSTemplateRequest_defaultSubstitutions :: Lens' SMSTemplateRequest (Maybe Text)

A JSON object that specifies the default values to use for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable. When you create a message that's based on the template, you can override these defaults with message-specific and address-specific variables and values.

sMSTemplateRequest_recommenderId :: Lens' SMSTemplateRequest (Maybe Text)

The unique identifier for the recommender model to use for the message template. Amazon Pinpoint uses this value to determine how to retrieve and process data from a recommender model when it sends messages that use the template, if the template contains message variables for recommendation data.

sMSTemplateRequest_tags :: Lens' SMSTemplateRequest (Maybe (HashMap Text Text))

A string-to-string map of key-value pairs that defines the tags to associate with the message template. Each tag consists of a required tag key and an associated tag value.