Network.AWS.Pinpoint.Types.SMSTemplateResponse

Description

Documentation

data SMSTemplateResponse

Provides information about the content and settings for a message template that can be used in text messages that are sent through the SMS channel.

See: newSMSTemplateResponse smart constructor.

Constructors

SMSTemplateResponse'

Fields

arn :: Maybe Text

The Amazon Resource Name (ARN) of the message template.

body :: Maybe Text

The message body that's used in text messages that are based on the message template.

templateDescription :: Maybe Text

The custom description of the message template.

defaultSubstitutions :: Maybe Text

The JSON object that specifies the default values that are used for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable.

version :: Maybe Text

The unique identifier, as an integer, for the active version of the message template, or the version of the template that you specified by using the version parameter in your request.

recommenderId :: Maybe Text

The unique identifier for the recommender model that's used by the message template.

tags :: Maybe (HashMap Text Text)

A string-to-string map of key-value pairs that identifies the tags that are associated with the message template. Each tag consists of a required tag key and an associated tag value.

lastModifiedDate :: Text

The date, in ISO 8601 format, when the message template was last modified.

creationDate :: Text

The date, in ISO 8601 format, when the message template was created.

templateName :: Text

The name of the message template.

templateType :: TemplateType

The type of channel that the message template is designed for. For an SMS template, this value is SMS.

newSMSTemplateResponse

Arguments

:: Text
$sel:lastModifiedDate:SMSTemplateResponse'

-> Text
$sel:creationDate:SMSTemplateResponse'

-> Text
$sel:templateName:SMSTemplateResponse'

-> TemplateType
$sel:templateType:SMSTemplateResponse'

-> SMSTemplateResponse

Create a value of SMSTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:arn:SMSTemplateResponse', sMSTemplateResponse_arn - The Amazon Resource Name (ARN) of the message template.

$sel:body:SMSTemplateResponse', sMSTemplateResponse_body - The message body that's used in text messages that are based on the message template.

$sel:templateDescription:SMSTemplateResponse', sMSTemplateResponse_templateDescription - The custom description of the message template.

$sel:defaultSubstitutions:SMSTemplateResponse', sMSTemplateResponse_defaultSubstitutions - The JSON object that specifies the default values that are used for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable.

$sel:version:SMSTemplateResponse', sMSTemplateResponse_version - The unique identifier, as an integer, for the active version of the message template, or the version of the template that you specified by using the version parameter in your request.

$sel:recommenderId:SMSTemplateResponse', sMSTemplateResponse_recommenderId - The unique identifier for the recommender model that's used by the message template.

$sel:tags:SMSTemplateResponse', sMSTemplateResponse_tags - A string-to-string map of key-value pairs that identifies the tags that are associated with the message template. Each tag consists of a required tag key and an associated tag value.

$sel:lastModifiedDate:SMSTemplateResponse', sMSTemplateResponse_lastModifiedDate - The date, in ISO 8601 format, when the message template was last modified.

$sel:creationDate:SMSTemplateResponse', sMSTemplateResponse_creationDate - The date, in ISO 8601 format, when the message template was created.

$sel:templateName:SMSTemplateResponse', sMSTemplateResponse_templateName - The name of the message template.

$sel:templateType:SMSTemplateResponse', sMSTemplateResponse_templateType - The type of channel that the message template is designed for. For an SMS template, this value is SMS.

sMSTemplateResponse_arn :: Lens' SMSTemplateResponse (Maybe Text)

The Amazon Resource Name (ARN) of the message template.

sMSTemplateResponse_body :: Lens' SMSTemplateResponse (Maybe Text)

The message body that's used in text messages that are based on the message template.

sMSTemplateResponse_templateDescription :: Lens' SMSTemplateResponse (Maybe Text)

The custom description of the message template.

sMSTemplateResponse_defaultSubstitutions :: Lens' SMSTemplateResponse (Maybe Text)

The JSON object that specifies the default values that are used for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable.

sMSTemplateResponse_version :: Lens' SMSTemplateResponse (Maybe Text)

The unique identifier, as an integer, for the active version of the message template, or the version of the template that you specified by using the version parameter in your request.

sMSTemplateResponse_recommenderId :: Lens' SMSTemplateResponse (Maybe Text)

The unique identifier for the recommender model that's used by the message template.

sMSTemplateResponse_tags :: Lens' SMSTemplateResponse (Maybe (HashMap Text Text))

A string-to-string map of key-value pairs that identifies the tags that are associated with the message template. Each tag consists of a required tag key and an associated tag value.

sMSTemplateResponse_lastModifiedDate :: Lens' SMSTemplateResponse Text

The date, in ISO 8601 format, when the message template was last modified.

sMSTemplateResponse_creationDate :: Lens' SMSTemplateResponse Text

The date, in ISO 8601 format, when the message template was created.

sMSTemplateResponse_templateName :: Lens' SMSTemplateResponse Text

The name of the message template.

sMSTemplateResponse_templateType :: Lens' SMSTemplateResponse TemplateType

The type of channel that the message template is designed for. For an SMS template, this value is SMS.