Network.AWS.Pinpoint.Types.Schedule

Description

Documentation

data Schedule

Specifies the schedule settings for a campaign.

See: newSchedule smart constructor.

Constructors

Schedule'

Fields

eventFilter :: Maybe CampaignEventFilter

The type of event that causes the campaign to be sent, if the value of the Frequency property is EVENT.

isLocalTime :: Maybe Bool

Specifies whether the start and end times for the campaign schedule use each recipient's local time. To base the schedule on each recipient's local time, set this value to true.

endTime :: Maybe Text

The scheduled time, in ISO 8601 format, when the campaign ended or will end.

frequency :: Maybe Frequency

Specifies how often the campaign is sent or whether the campaign is sent in response to a specific event.

quietTime :: Maybe QuietTime

The default quiet time for the campaign. Quiet time is a specific time range when a campaign doesn't send messages to endpoints, if all the following conditions are met:

The EndpointDemographic.Timezone property of the endpoint is set to a valid value.

The current time in the endpoint's time zone is later than or equal to the time specified by the QuietTime.Start property for the campaign.

The current time in the endpoint's time zone is earlier than or equal to the time specified by the QuietTime.End property for the campaign.

If any of the preceding conditions isn't met, the endpoint will receive messages from the campaign, even if quiet time is enabled.

timezone :: Maybe Text

The starting UTC offset for the campaign schedule, if the value of the IsLocalTime property is true. Valid values are: UTC, UTC+01, UTC+02, UTC+03, UTC+03:30, UTC+04, UTC+04:30, UTC+05, UTC+05:30, UTC+05:45, UTC+06, UTC+06:30, UTC+07, UTC+08, UTC+09, UTC+09:30, UTC+10, UTC+10:30, UTC+11, UTC+12, UTC+13, UTC-02, UTC-03, UTC-04, UTC-05, UTC-06, UTC-07, UTC-08, UTC-09, UTC-10, and UTC-11.

startTime :: Text

The scheduled time when the campaign began or will begin. Valid values are: IMMEDIATE, to start the campaign immediately; or, a specific time in ISO 8601 format.

newSchedule

Arguments

:: Text
$sel:startTime:Schedule'

-> Schedule

Create a value of Schedule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventFilter:Schedule', schedule_eventFilter - The type of event that causes the campaign to be sent, if the value of the Frequency property is EVENT.

$sel:isLocalTime:Schedule', schedule_isLocalTime - Specifies whether the start and end times for the campaign schedule use each recipient's local time. To base the schedule on each recipient's local time, set this value to true.

$sel:endTime:Schedule', schedule_endTime - The scheduled time, in ISO 8601 format, when the campaign ended or will end.

$sel:frequency:Schedule', schedule_frequency - Specifies how often the campaign is sent or whether the campaign is sent in response to a specific event.

$sel:quietTime:Schedule', schedule_quietTime - The default quiet time for the campaign. Quiet time is a specific time range when a campaign doesn't send messages to endpoints, if all the following conditions are met:

The EndpointDemographic.Timezone property of the endpoint is set to a valid value.

The current time in the endpoint's time zone is later than or equal to the time specified by the QuietTime.Start property for the campaign.

The current time in the endpoint's time zone is earlier than or equal to the time specified by the QuietTime.End property for the campaign.

If any of the preceding conditions isn't met, the endpoint will receive messages from the campaign, even if quiet time is enabled.

$sel:timezone:Schedule', schedule_timezone - The starting UTC offset for the campaign schedule, if the value of the IsLocalTime property is true. Valid values are: UTC, UTC+01, UTC+02, UTC+03, UTC+03:30, UTC+04, UTC+04:30, UTC+05, UTC+05:30, UTC+05:45, UTC+06, UTC+06:30, UTC+07, UTC+08, UTC+09, UTC+09:30, UTC+10, UTC+10:30, UTC+11, UTC+12, UTC+13, UTC-02, UTC-03, UTC-04, UTC-05, UTC-06, UTC-07, UTC-08, UTC-09, UTC-10, and UTC-11.

$sel:startTime:Schedule', schedule_startTime - The scheduled time when the campaign began or will begin. Valid values are: IMMEDIATE, to start the campaign immediately; or, a specific time in ISO 8601 format.

schedule_eventFilter :: Lens' Schedule (Maybe CampaignEventFilter)

The type of event that causes the campaign to be sent, if the value of the Frequency property is EVENT.

schedule_isLocalTime :: Lens' Schedule (Maybe Bool)

Specifies whether the start and end times for the campaign schedule use each recipient's local time. To base the schedule on each recipient's local time, set this value to true.

schedule_endTime :: Lens' Schedule (Maybe Text)

The scheduled time, in ISO 8601 format, when the campaign ended or will end.

schedule_frequency :: Lens' Schedule (Maybe Frequency)

Specifies how often the campaign is sent or whether the campaign is sent in response to a specific event.

schedule_quietTime :: Lens' Schedule (Maybe QuietTime)

The default quiet time for the campaign. Quiet time is a specific time range when a campaign doesn't send messages to endpoints, if all the following conditions are met:

The EndpointDemographic.Timezone property of the endpoint is set to a valid value.

The current time in the endpoint's time zone is later than or equal to the time specified by the QuietTime.Start property for the campaign.

The current time in the endpoint's time zone is earlier than or equal to the time specified by the QuietTime.End property for the campaign.

If any of the preceding conditions isn't met, the endpoint will receive messages from the campaign, even if quiet time is enabled.

schedule_timezone :: Lens' Schedule (Maybe Text)

The starting UTC offset for the campaign schedule, if the value of the IsLocalTime property is true. Valid values are: UTC, UTC+01, UTC+02, UTC+03, UTC+03:30, UTC+04, UTC+04:30, UTC+05, UTC+05:30, UTC+05:45, UTC+06, UTC+06:30, UTC+07, UTC+08, UTC+09, UTC+09:30, UTC+10, UTC+10:30, UTC+11, UTC+12, UTC+13, UTC-02, UTC-03, UTC-04, UTC-05, UTC-06, UTC-07, UTC-08, UTC-09, UTC-10, and UTC-11.

schedule_startTime :: Lens' Schedule Text

The scheduled time when the campaign began or will begin. Valid values are: IMMEDIATE, to start the campaign immediately; or, a specific time in ISO 8601 format.