Network.AWS.Pinpoint.Types.SegmentCondition

Description

Documentation

data SegmentCondition

Specifies a segment to associate with an activity in a journey.

See: newSegmentCondition smart constructor.

Constructors

SegmentCondition'

Fields

segmentId :: Text

The unique identifier for the segment to associate with the activity.

newSegmentCondition

Arguments

:: Text
$sel:segmentId:SegmentCondition'

-> SegmentCondition

Create a value of SegmentCondition with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:segmentId:SegmentCondition', segmentCondition_segmentId - The unique identifier for the segment to associate with the activity.

segmentCondition_segmentId :: Lens' SegmentCondition Text

The unique identifier for the segment to associate with the activity.