Network.AWS.Pinpoint.Types.SegmentGroup

Description

Documentation

data SegmentGroup

Specifies the base segments and dimensions for a segment, and the relationships between these base segments and dimensions.

See: newSegmentGroup smart constructor.

Constructors

SegmentGroup'

Fields

dimensions :: Maybe [SegmentDimensions]

An array that defines the dimensions for the segment.

type' :: Maybe Type

Specifies how to handle multiple dimensions for the segment. For example, if you specify three dimensions for the segment, whether the resulting segment includes endpoints that match all, any, or none of the dimensions.

sourceType :: Maybe SourceType

Specifies how to handle multiple base segments for the segment. For example, if you specify three base segments for the segment, whether the resulting segment is based on all, any, or none of the base segments.

sourceSegments :: Maybe [SegmentReference]

The base segment to build the segment on. A base segment, also referred to as a source segment, defines the initial population of endpoints for a segment. When you add dimensions to a segment, Amazon Pinpoint filters the base segment by using the dimensions that you specify.

You can specify more than one dimensional segment or only one imported segment. If you specify an imported segment, the Amazon Pinpoint console displays a segment size estimate that indicates the size of the imported segment without any filters applied to it.

newSegmentGroup :: SegmentGroup

Create a value of SegmentGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dimensions:SegmentGroup', segmentGroup_dimensions - An array that defines the dimensions for the segment.

$sel:type':SegmentGroup', segmentGroup_type - Specifies how to handle multiple dimensions for the segment. For example, if you specify three dimensions for the segment, whether the resulting segment includes endpoints that match all, any, or none of the dimensions.

$sel:sourceType:SegmentGroup', segmentGroup_sourceType - Specifies how to handle multiple base segments for the segment. For example, if you specify three base segments for the segment, whether the resulting segment is based on all, any, or none of the base segments.

$sel:sourceSegments:SegmentGroup', segmentGroup_sourceSegments - The base segment to build the segment on. A base segment, also referred to as a source segment, defines the initial population of endpoints for a segment. When you add dimensions to a segment, Amazon Pinpoint filters the base segment by using the dimensions that you specify.

You can specify more than one dimensional segment or only one imported segment. If you specify an imported segment, the Amazon Pinpoint console displays a segment size estimate that indicates the size of the imported segment without any filters applied to it.

segmentGroup_dimensions :: Lens' SegmentGroup (Maybe [SegmentDimensions])

An array that defines the dimensions for the segment.

segmentGroup_type :: Lens' SegmentGroup (Maybe Type)

Specifies how to handle multiple dimensions for the segment. For example, if you specify three dimensions for the segment, whether the resulting segment includes endpoints that match all, any, or none of the dimensions.

segmentGroup_sourceType :: Lens' SegmentGroup (Maybe SourceType)

Specifies how to handle multiple base segments for the segment. For example, if you specify three base segments for the segment, whether the resulting segment is based on all, any, or none of the base segments.

segmentGroup_sourceSegments :: Lens' SegmentGroup (Maybe [SegmentReference])

The base segment to build the segment on. A base segment, also referred to as a source segment, defines the initial population of endpoints for a segment. When you add dimensions to a segment, Amazon Pinpoint filters the base segment by using the dimensions that you specify.

You can specify more than one dimensional segment or only one imported segment. If you specify an imported segment, the Amazon Pinpoint console displays a segment size estimate that indicates the size of the imported segment without any filters applied to it.