Network.AWS.Pinpoint.Types.SegmentImportResource

Description

Documentation

data SegmentImportResource

Provides information about the import job that created a segment. An import job is a job that creates a user segment by importing endpoint definitions.

See: newSegmentImportResource smart constructor.

Constructors

SegmentImportResource'

Fields

channelCounts :: Maybe (HashMap Text Int)

The number of channel types in the endpoint definitions that were imported to create the segment.

format :: DefinitionFormat

The format of the files that were imported to create the segment. Valid values are: CSV, for comma-separated values format; and, JSON, for newline-delimited JSON format.

s3Url :: Text

The URL of the Amazon Simple Storage Service (Amazon S3) bucket that the endpoint definitions were imported from to create the segment.

size :: Int

The number of endpoint definitions that were imported successfully to create the segment.

externalId :: Text

(Deprecated) Your AWS account ID, which you assigned to an external ID key in an IAM trust policy. Amazon Pinpoint previously used this value to assume an IAM role when importing endpoint definitions, but we removed this requirement. We don't recommend use of external IDs for IAM roles that are assumed by Amazon Pinpoint.

roleArn :: Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that authorized Amazon Pinpoint to access the Amazon S3 location to import endpoint definitions from.

newSegmentImportResource

Arguments

:: DefinitionFormat
$sel:format:SegmentImportResource'

-> Text
$sel:s3Url:SegmentImportResource'

-> Int
$sel:size:SegmentImportResource'

-> Text
$sel:externalId:SegmentImportResource'

-> Text
$sel:roleArn:SegmentImportResource'

-> SegmentImportResource

Create a value of SegmentImportResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:channelCounts:SegmentImportResource', segmentImportResource_channelCounts - The number of channel types in the endpoint definitions that were imported to create the segment.

$sel:format:SegmentImportResource', segmentImportResource_format - The format of the files that were imported to create the segment. Valid values are: CSV, for comma-separated values format; and, JSON, for newline-delimited JSON format.

$sel:s3Url:SegmentImportResource', segmentImportResource_s3Url - The URL of the Amazon Simple Storage Service (Amazon S3) bucket that the endpoint definitions were imported from to create the segment.

$sel:size:SegmentImportResource', segmentImportResource_size - The number of endpoint definitions that were imported successfully to create the segment.

$sel:externalId:SegmentImportResource', segmentImportResource_externalId - (Deprecated) Your AWS account ID, which you assigned to an external ID key in an IAM trust policy. Amazon Pinpoint previously used this value to assume an IAM role when importing endpoint definitions, but we removed this requirement. We don't recommend use of external IDs for IAM roles that are assumed by Amazon Pinpoint.

$sel:roleArn:SegmentImportResource', segmentImportResource_roleArn - The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that authorized Amazon Pinpoint to access the Amazon S3 location to import endpoint definitions from.

segmentImportResource_channelCounts :: Lens' SegmentImportResource (Maybe (HashMap Text Int))

The number of channel types in the endpoint definitions that were imported to create the segment.

segmentImportResource_format :: Lens' SegmentImportResource DefinitionFormat

The format of the files that were imported to create the segment. Valid values are: CSV, for comma-separated values format; and, JSON, for newline-delimited JSON format.

segmentImportResource_s3Url :: Lens' SegmentImportResource Text

The URL of the Amazon Simple Storage Service (Amazon S3) bucket that the endpoint definitions were imported from to create the segment.

segmentImportResource_size :: Lens' SegmentImportResource Int

The number of endpoint definitions that were imported successfully to create the segment.

segmentImportResource_externalId :: Lens' SegmentImportResource Text

(Deprecated) Your AWS account ID, which you assigned to an external ID key in an IAM trust policy. Amazon Pinpoint previously used this value to assume an IAM role when importing endpoint definitions, but we removed this requirement. We don't recommend use of external IDs for IAM roles that are assumed by Amazon Pinpoint.

segmentImportResource_roleArn :: Lens' SegmentImportResource Text

The Amazon Resource Name (ARN) of the AWS Identity and Access Management (IAM) role that authorized Amazon Pinpoint to access the Amazon S3 location to import endpoint definitions from.