Network.AWS.Pinpoint.Types.SegmentResponse

Description

Documentation

data SegmentResponse

Provides information about the configuration, dimension, and other settings for a segment.

See: newSegmentResponse smart constructor.

Constructors

SegmentResponse'

Fields

lastModifiedDate :: Maybe Text

The date and time when the segment was last modified.

segmentGroups :: Maybe SegmentGroupList

A list of one or more segment groups that apply to the segment. Each segment group consists of zero or more base segments and the dimensions that are applied to those base segments.

name :: Maybe Text

The name of the segment.

version :: Maybe Int

The version number of the segment.

importDefinition :: Maybe SegmentImportResource

The settings for the import job that's associated with the segment.

dimensions :: Maybe SegmentDimensions

The dimension settings for the segment.

tags :: Maybe (HashMap Text Text)

A string-to-string map of key-value pairs that identifies the tags that are associated with the segment. Each tag consists of a required tag key and an associated tag value.

segmentType :: SegmentType

The segment type. Valid values are:

DIMENSIONAL - A dynamic segment, which is a segment that uses selection criteria that you specify and is based on endpoint data that's reported by your app. Dynamic segments can change over time.

IMPORT - A static segment, which is a segment that uses selection criteria that you specify and is based on endpoint definitions that you import from a file. Imported segments are static; they don't change over time.

creationDate :: Text

The date and time when the segment was created.

id :: Text

The unique identifier for the segment.

arn :: Text

The Amazon Resource Name (ARN) of the segment.

applicationId :: Text

The unique identifier for the application that the segment is associated with.

newSegmentResponse

Arguments

:: SegmentType
$sel:segmentType:SegmentResponse'

-> Text
$sel:creationDate:SegmentResponse'

-> Text
$sel:id:SegmentResponse'

-> Text
$sel:arn:SegmentResponse'

-> Text
$sel:applicationId:SegmentResponse'

-> SegmentResponse

Create a value of SegmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:SegmentResponse', segmentResponse_lastModifiedDate - The date and time when the segment was last modified.

$sel:segmentGroups:SegmentResponse', segmentResponse_segmentGroups - A list of one or more segment groups that apply to the segment. Each segment group consists of zero or more base segments and the dimensions that are applied to those base segments.

$sel:name:SegmentResponse', segmentResponse_name - The name of the segment.

$sel:version:SegmentResponse', segmentResponse_version - The version number of the segment.

$sel:importDefinition:SegmentResponse', segmentResponse_importDefinition - The settings for the import job that's associated with the segment.

$sel:dimensions:SegmentResponse', segmentResponse_dimensions - The dimension settings for the segment.

$sel:tags:SegmentResponse', segmentResponse_tags - A string-to-string map of key-value pairs that identifies the tags that are associated with the segment. Each tag consists of a required tag key and an associated tag value.

$sel:segmentType:SegmentResponse', segmentResponse_segmentType - The segment type. Valid values are:

DIMENSIONAL - A dynamic segment, which is a segment that uses selection criteria that you specify and is based on endpoint data that's reported by your app. Dynamic segments can change over time.

IMPORT - A static segment, which is a segment that uses selection criteria that you specify and is based on endpoint definitions that you import from a file. Imported segments are static; they don't change over time.

$sel:creationDate:SegmentResponse', segmentResponse_creationDate - The date and time when the segment was created.

$sel:id:SegmentResponse', segmentResponse_id - The unique identifier for the segment.

$sel:arn:SegmentResponse', segmentResponse_arn - The Amazon Resource Name (ARN) of the segment.

$sel:applicationId:SegmentResponse', segmentResponse_applicationId - The unique identifier for the application that the segment is associated with.

segmentResponse_lastModifiedDate :: Lens' SegmentResponse (Maybe Text)

The date and time when the segment was last modified.

segmentResponse_segmentGroups :: Lens' SegmentResponse (Maybe SegmentGroupList)

A list of one or more segment groups that apply to the segment. Each segment group consists of zero or more base segments and the dimensions that are applied to those base segments.

segmentResponse_name :: Lens' SegmentResponse (Maybe Text)

The name of the segment.

segmentResponse_version :: Lens' SegmentResponse (Maybe Int)

The version number of the segment.

segmentResponse_importDefinition :: Lens' SegmentResponse (Maybe SegmentImportResource)

The settings for the import job that's associated with the segment.

segmentResponse_dimensions :: Lens' SegmentResponse (Maybe SegmentDimensions)

The dimension settings for the segment.

segmentResponse_tags :: Lens' SegmentResponse (Maybe (HashMap Text Text))

A string-to-string map of key-value pairs that identifies the tags that are associated with the segment. Each tag consists of a required tag key and an associated tag value.

segmentResponse_segmentType :: Lens' SegmentResponse SegmentType

The segment type. Valid values are:

DIMENSIONAL - A dynamic segment, which is a segment that uses selection criteria that you specify and is based on endpoint data that's reported by your app. Dynamic segments can change over time.

IMPORT - A static segment, which is a segment that uses selection criteria that you specify and is based on endpoint definitions that you import from a file. Imported segments are static; they don't change over time.

segmentResponse_creationDate :: Lens' SegmentResponse Text

The date and time when the segment was created.

segmentResponse_id :: Lens' SegmentResponse Text

The unique identifier for the segment.

segmentResponse_arn :: Lens' SegmentResponse Text

The Amazon Resource Name (ARN) of the segment.

segmentResponse_applicationId :: Lens' SegmentResponse Text

The unique identifier for the application that the segment is associated with.