Network.AWS.Pinpoint.Types.SendUsersMessageRequest

Description

Documentation

data SendUsersMessageRequest

Specifies the configuration and other settings for a message to send to all the endpoints that are associated with a list of users.

See: newSendUsersMessageRequest smart constructor.

Constructors

SendUsersMessageRequest'

Fields

context :: Maybe (HashMap Text Text)

A map of custom attribute-value pairs. For a push notification, Amazon Pinpoint adds these attributes to the data.pinpoint object in the body of the notification payload. Amazon Pinpoint also provides these attributes in the events that it generates for users-messages deliveries.

traceId :: Maybe Text

The unique identifier for tracing the message. This identifier is visible to message recipients.

templateConfiguration :: Maybe TemplateConfiguration

The message template to use for the message.

messageConfiguration :: DirectMessageConfiguration

The settings and content for the default message and any default messages that you defined for specific channels.

users :: HashMap Text EndpointSendConfiguration

A map that associates user IDs with EndpointSendConfiguration objects. You can use an EndpointSendConfiguration object to tailor the message for a user by specifying settings such as content overrides and message variables.

newSendUsersMessageRequest

Arguments

:: DirectMessageConfiguration
$sel:messageConfiguration:SendUsersMessageRequest'

-> SendUsersMessageRequest

Create a value of SendUsersMessageRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:context:SendUsersMessageRequest', sendUsersMessageRequest_context - A map of custom attribute-value pairs. For a push notification, Amazon Pinpoint adds these attributes to the data.pinpoint object in the body of the notification payload. Amazon Pinpoint also provides these attributes in the events that it generates for users-messages deliveries.

$sel:traceId:SendUsersMessageRequest', sendUsersMessageRequest_traceId - The unique identifier for tracing the message. This identifier is visible to message recipients.

$sel:templateConfiguration:SendUsersMessageRequest', sendUsersMessageRequest_templateConfiguration - The message template to use for the message.

$sel:messageConfiguration:SendUsersMessageRequest', sendUsersMessageRequest_messageConfiguration - The settings and content for the default message and any default messages that you defined for specific channels.

$sel:users:SendUsersMessageRequest', sendUsersMessageRequest_users - A map that associates user IDs with EndpointSendConfiguration objects. You can use an EndpointSendConfiguration object to tailor the message for a user by specifying settings such as content overrides and message variables.

sendUsersMessageRequest_context :: Lens' SendUsersMessageRequest (Maybe (HashMap Text Text))

A map of custom attribute-value pairs. For a push notification, Amazon Pinpoint adds these attributes to the data.pinpoint object in the body of the notification payload. Amazon Pinpoint also provides these attributes in the events that it generates for users-messages deliveries.

sendUsersMessageRequest_traceId :: Lens' SendUsersMessageRequest (Maybe Text)

The unique identifier for tracing the message. This identifier is visible to message recipients.

sendUsersMessageRequest_templateConfiguration :: Lens' SendUsersMessageRequest (Maybe TemplateConfiguration)

The message template to use for the message.

sendUsersMessageRequest_messageConfiguration :: Lens' SendUsersMessageRequest DirectMessageConfiguration

The settings and content for the default message and any default messages that you defined for specific channels.

sendUsersMessageRequest_users :: Lens' SendUsersMessageRequest (HashMap Text EndpointSendConfiguration)

A map that associates user IDs with EndpointSendConfiguration objects. You can use an EndpointSendConfiguration object to tailor the message for a user by specifying settings such as content overrides and message variables.