Network.AWS.Pinpoint.Types.TemplateConfiguration

Description

Documentation

data TemplateConfiguration

Specifies the message template to use for the message, for each type of channel.

See: newTemplateConfiguration smart constructor.

Constructors

TemplateConfiguration'

Fields

emailTemplate :: Maybe Template

The email template to use for the message.

voiceTemplate :: Maybe Template

The voice template to use for the message. This object isn't supported for campaigns.

sMSTemplate :: Maybe Template

The SMS template to use for the message.

pushTemplate :: Maybe Template

The push notification template to use for the message.

newTemplateConfiguration :: TemplateConfiguration

Create a value of TemplateConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:emailTemplate:TemplateConfiguration', templateConfiguration_emailTemplate - The email template to use for the message.

$sel:voiceTemplate:TemplateConfiguration', templateConfiguration_voiceTemplate - The voice template to use for the message. This object isn't supported for campaigns.

$sel:sMSTemplate:TemplateConfiguration', templateConfiguration_sMSTemplate - The SMS template to use for the message.

$sel:pushTemplate:TemplateConfiguration', templateConfiguration_pushTemplate - The push notification template to use for the message.

templateConfiguration_emailTemplate :: Lens' TemplateConfiguration (Maybe Template)

The email template to use for the message.

templateConfiguration_voiceTemplate :: Lens' TemplateConfiguration (Maybe Template)

The voice template to use for the message. This object isn't supported for campaigns.

templateConfiguration_sMSTemplate :: Lens' TemplateConfiguration (Maybe Template)

The SMS template to use for the message.

templateConfiguration_pushTemplate :: Lens' TemplateConfiguration (Maybe Template)

The push notification template to use for the message.