Network.AWS.Pinpoint.Types.TemplateResponse

Description

Documentation

data TemplateResponse

Provides information about a message template that's associated with your Amazon Pinpoint account.

See: newTemplateResponse smart constructor.

Constructors

TemplateResponse'

Fields

templateDescription :: Maybe Text

The custom description of the message template. This value isn't included in a TemplateResponse object. To retrieve the description of a template, use the GetEmailTemplate, GetPushTemplate, GetSmsTemplate, or GetVoiceTemplate operation, depending on the type of template that you want to retrieve the description for.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the message template. This value isn't included in a TemplateResponse object. To retrieve the ARN of a template, use the GetEmailTemplate, GetPushTemplate, GetSmsTemplate, or GetVoiceTemplate operation, depending on the type of template that you want to retrieve the ARN for.

version :: Maybe Text

The unique identifier, as an integer, for the active version of the message template.

defaultSubstitutions :: Maybe Text

The JSON object that specifies the default values that are used for message variables in the message template. This object isn't included in a TemplateResponse object. To retrieve this object for a template, use the GetEmailTemplate, GetPushTemplate, GetSmsTemplate, or GetVoiceTemplate operation, depending on the type of template that you want to retrieve the object for.

tags :: Maybe (HashMap Text Text)

A map of key-value pairs that identifies the tags that are associated with the message template. This object isn't included in a TemplateResponse object. To retrieve this object for a template, use the GetEmailTemplate, GetPushTemplate, GetSmsTemplate, or GetVoiceTemplate operation, depending on the type of template that you want to retrieve the object for.

lastModifiedDate :: Text

The date, in ISO 8601 format, when the message template was last modified.

creationDate :: Text

The date, in ISO 8601 format, when the message template was created.

templateName :: Text

The name of the message template.

templateType :: TemplateType

The type of channel that the message template is designed for. Possible values are: EMAIL, PUSH, SMS, and VOICE.

newTemplateResponse

Arguments

:: Text
$sel:lastModifiedDate:TemplateResponse'

-> Text
$sel:creationDate:TemplateResponse'

-> Text
$sel:templateName:TemplateResponse'

-> TemplateType
$sel:templateType:TemplateResponse'

-> TemplateResponse

Create a value of TemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateDescription:TemplateResponse', templateResponse_templateDescription - The custom description of the message template. This value isn't included in a TemplateResponse object. To retrieve the description of a template, use the GetEmailTemplate, GetPushTemplate, GetSmsTemplate, or GetVoiceTemplate operation, depending on the type of template that you want to retrieve the description for.

$sel:arn:TemplateResponse', templateResponse_arn - The Amazon Resource Name (ARN) of the message template. This value isn't included in a TemplateResponse object. To retrieve the ARN of a template, use the GetEmailTemplate, GetPushTemplate, GetSmsTemplate, or GetVoiceTemplate operation, depending on the type of template that you want to retrieve the ARN for.

$sel:version:TemplateResponse', templateResponse_version - The unique identifier, as an integer, for the active version of the message template.

$sel:defaultSubstitutions:TemplateResponse', templateResponse_defaultSubstitutions - The JSON object that specifies the default values that are used for message variables in the message template. This object isn't included in a TemplateResponse object. To retrieve this object for a template, use the GetEmailTemplate, GetPushTemplate, GetSmsTemplate, or GetVoiceTemplate operation, depending on the type of template that you want to retrieve the object for.

$sel:tags:TemplateResponse', templateResponse_tags - A map of key-value pairs that identifies the tags that are associated with the message template. This object isn't included in a TemplateResponse object. To retrieve this object for a template, use the GetEmailTemplate, GetPushTemplate, GetSmsTemplate, or GetVoiceTemplate operation, depending on the type of template that you want to retrieve the object for.

$sel:lastModifiedDate:TemplateResponse', templateResponse_lastModifiedDate - The date, in ISO 8601 format, when the message template was last modified.

$sel:creationDate:TemplateResponse', templateResponse_creationDate - The date, in ISO 8601 format, when the message template was created.

$sel:templateName:TemplateResponse', templateResponse_templateName - The name of the message template.

$sel:templateType:TemplateResponse', templateResponse_templateType - The type of channel that the message template is designed for. Possible values are: EMAIL, PUSH, SMS, and VOICE.

templateResponse_templateDescription :: Lens' TemplateResponse (Maybe Text)

The custom description of the message template. This value isn't included in a TemplateResponse object. To retrieve the description of a template, use the GetEmailTemplate, GetPushTemplate, GetSmsTemplate, or GetVoiceTemplate operation, depending on the type of template that you want to retrieve the description for.

templateResponse_arn :: Lens' TemplateResponse (Maybe Text)

The Amazon Resource Name (ARN) of the message template. This value isn't included in a TemplateResponse object. To retrieve the ARN of a template, use the GetEmailTemplate, GetPushTemplate, GetSmsTemplate, or GetVoiceTemplate operation, depending on the type of template that you want to retrieve the ARN for.

templateResponse_version :: Lens' TemplateResponse (Maybe Text)

The unique identifier, as an integer, for the active version of the message template.

templateResponse_defaultSubstitutions :: Lens' TemplateResponse (Maybe Text)

The JSON object that specifies the default values that are used for message variables in the message template. This object isn't included in a TemplateResponse object. To retrieve this object for a template, use the GetEmailTemplate, GetPushTemplate, GetSmsTemplate, or GetVoiceTemplate operation, depending on the type of template that you want to retrieve the object for.

templateResponse_tags :: Lens' TemplateResponse (Maybe (HashMap Text Text))

A map of key-value pairs that identifies the tags that are associated with the message template. This object isn't included in a TemplateResponse object. To retrieve this object for a template, use the GetEmailTemplate, GetPushTemplate, GetSmsTemplate, or GetVoiceTemplate operation, depending on the type of template that you want to retrieve the object for.

templateResponse_lastModifiedDate :: Lens' TemplateResponse Text

The date, in ISO 8601 format, when the message template was last modified.

templateResponse_creationDate :: Lens' TemplateResponse Text

The date, in ISO 8601 format, when the message template was created.

templateResponse_templateName :: Lens' TemplateResponse Text

The name of the message template.

templateResponse_templateType :: Lens' TemplateResponse TemplateType

The type of channel that the message template is designed for. Possible values are: EMAIL, PUSH, SMS, and VOICE.