Network.AWS.Pinpoint.Types.TemplateVersionResponse

Description

Documentation

data TemplateVersionResponse

Provides information about a specific version of a message template.

See: newTemplateVersionResponse smart constructor.

Constructors

TemplateVersionResponse'

Fields

templateDescription :: Maybe Text

The custom description of the version of the message template.

version :: Maybe Text

The unique identifier for the version of the message template. This value is an integer that Amazon Pinpoint automatically increments and assigns to each new version of a template.

defaultSubstitutions :: Maybe Text

A JSON object that specifies the default values that are used for message variables in the version of the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable.

lastModifiedDate :: Text

The date, in ISO 8601 format, when the version of the message template was last modified.

creationDate :: Text

The date, in ISO 8601 format, when the version of the message template was created.

templateName :: Text

The name of the message template.

templateType :: Text

The type of channel that the message template is designed for. Possible values are: EMAIL, PUSH, SMS, and VOICE.

newTemplateVersionResponse

Arguments

:: Text
$sel:lastModifiedDate:TemplateVersionResponse'

-> Text
$sel:creationDate:TemplateVersionResponse'

-> Text
$sel:templateName:TemplateVersionResponse'

-> Text
$sel:templateType:TemplateVersionResponse'

-> TemplateVersionResponse

Create a value of TemplateVersionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateDescription:TemplateVersionResponse', templateVersionResponse_templateDescription - The custom description of the version of the message template.

$sel:version:TemplateVersionResponse', templateVersionResponse_version - The unique identifier for the version of the message template. This value is an integer that Amazon Pinpoint automatically increments and assigns to each new version of a template.

$sel:defaultSubstitutions:TemplateVersionResponse', templateVersionResponse_defaultSubstitutions - A JSON object that specifies the default values that are used for message variables in the version of the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable.

$sel:lastModifiedDate:TemplateVersionResponse', templateVersionResponse_lastModifiedDate - The date, in ISO 8601 format, when the version of the message template was last modified.

$sel:creationDate:TemplateVersionResponse', templateVersionResponse_creationDate - The date, in ISO 8601 format, when the version of the message template was created.

$sel:templateName:TemplateVersionResponse', templateVersionResponse_templateName - The name of the message template.

$sel:templateType:TemplateVersionResponse', templateVersionResponse_templateType - The type of channel that the message template is designed for. Possible values are: EMAIL, PUSH, SMS, and VOICE.

templateVersionResponse_templateDescription :: Lens' TemplateVersionResponse (Maybe Text)

The custom description of the version of the message template.

templateVersionResponse_version :: Lens' TemplateVersionResponse (Maybe Text)

The unique identifier for the version of the message template. This value is an integer that Amazon Pinpoint automatically increments and assigns to each new version of a template.

templateVersionResponse_defaultSubstitutions :: Lens' TemplateVersionResponse (Maybe Text)

A JSON object that specifies the default values that are used for message variables in the version of the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable.

templateVersionResponse_lastModifiedDate :: Lens' TemplateVersionResponse Text

The date, in ISO 8601 format, when the version of the message template was last modified.

templateVersionResponse_creationDate :: Lens' TemplateVersionResponse Text

The date, in ISO 8601 format, when the version of the message template was created.

templateVersionResponse_templateName :: Lens' TemplateVersionResponse Text

The name of the message template.

templateVersionResponse_templateType :: Lens' TemplateVersionResponse Text

The type of channel that the message template is designed for. Possible values are: EMAIL, PUSH, SMS, and VOICE.