Network.AWS.Pinpoint.Types.TreatmentResource

Description

Documentation

data TreatmentResource

Specifies the settings for a campaign treatment. A treatment is a variation of a campaign that's used for A/B testing of a campaign.

See: newTreatmentResource smart constructor.

Constructors

TreatmentResource'

Fields

customDeliveryConfiguration :: Maybe CustomDeliveryConfiguration

The delivery configuration settings for sending the treatment through a custom channel. This object is required if the MessageConfiguration object for the treatment specifies a CustomMessage object.

state :: Maybe CampaignState

The current status of the treatment.

schedule :: Maybe Schedule

The schedule settings for the treatment.

templateConfiguration :: Maybe TemplateConfiguration

The message template to use for the treatment.

treatmentName :: Maybe Text

The custom name of the treatment.

treatmentDescription :: Maybe Text

The custom description of the treatment.

messageConfiguration :: Maybe MessageConfiguration

The message configuration settings for the treatment.

id :: Text

The unique identifier for the treatment.

sizePercent :: Int

The allocated percentage of users (segment members) that the treatment is sent to.

newTreatmentResource

Arguments

:: Text
$sel:id:TreatmentResource'

-> Int
$sel:sizePercent:TreatmentResource'

-> TreatmentResource

Create a value of TreatmentResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customDeliveryConfiguration:TreatmentResource', treatmentResource_customDeliveryConfiguration - The delivery configuration settings for sending the treatment through a custom channel. This object is required if the MessageConfiguration object for the treatment specifies a CustomMessage object.

$sel:state:TreatmentResource', treatmentResource_state - The current status of the treatment.

$sel:schedule:TreatmentResource', treatmentResource_schedule - The schedule settings for the treatment.

$sel:templateConfiguration:TreatmentResource', treatmentResource_templateConfiguration - The message template to use for the treatment.

$sel:treatmentName:TreatmentResource', treatmentResource_treatmentName - The custom name of the treatment.

$sel:treatmentDescription:TreatmentResource', treatmentResource_treatmentDescription - The custom description of the treatment.

$sel:messageConfiguration:TreatmentResource', treatmentResource_messageConfiguration - The message configuration settings for the treatment.

$sel:id:TreatmentResource', treatmentResource_id - The unique identifier for the treatment.

$sel:sizePercent:TreatmentResource', treatmentResource_sizePercent - The allocated percentage of users (segment members) that the treatment is sent to.

treatmentResource_customDeliveryConfiguration :: Lens' TreatmentResource (Maybe CustomDeliveryConfiguration)

The delivery configuration settings for sending the treatment through a custom channel. This object is required if the MessageConfiguration object for the treatment specifies a CustomMessage object.

treatmentResource_state :: Lens' TreatmentResource (Maybe CampaignState)

The current status of the treatment.

treatmentResource_schedule :: Lens' TreatmentResource (Maybe Schedule)

The schedule settings for the treatment.

treatmentResource_templateConfiguration :: Lens' TreatmentResource (Maybe TemplateConfiguration)

The message template to use for the treatment.

treatmentResource_treatmentName :: Lens' TreatmentResource (Maybe Text)

The custom name of the treatment.

treatmentResource_treatmentDescription :: Lens' TreatmentResource (Maybe Text)

The custom description of the treatment.

treatmentResource_messageConfiguration :: Lens' TreatmentResource (Maybe MessageConfiguration)

The message configuration settings for the treatment.

treatmentResource_id :: Lens' TreatmentResource Text

The unique identifier for the treatment.

treatmentResource_sizePercent :: Lens' TreatmentResource Int

The allocated percentage of users (segment members) that the treatment is sent to.