Network.AWS.Pinpoint.Types.VoiceChannelResponse

Description

Documentation

data VoiceChannelResponse

Provides information about the status and settings of the voice channel for an application.

See: newVoiceChannelResponse smart constructor.

Constructors

VoiceChannelResponse'

Fields

lastModifiedDate :: Maybe Text

The date and time, in ISO 8601 format, when the voice channel was last modified.

enabled :: Maybe Bool

Specifies whether the voice channel is enabled for the application.

isArchived :: Maybe Bool

Specifies whether the voice channel is archived.

applicationId :: Maybe Text

The unique identifier for the application that the voice channel applies to.

version :: Maybe Int

The current version of the voice channel.

id :: Maybe Text

(Deprecated) An identifier for the voice channel. This property is retained only for backward compatibility.

creationDate :: Maybe Text

The date and time, in ISO 8601 format, when the voice channel was enabled.

lastModifiedBy :: Maybe Text

The user who last modified the voice channel.

hasCredential :: Maybe Bool

(Not used) This property is retained only for backward compatibility.

platform :: Text

The type of messaging or notification platform for the channel. For the voice channel, this value is VOICE.

newVoiceChannelResponse

Arguments

:: Text
$sel:platform:VoiceChannelResponse'

-> VoiceChannelResponse

Create a value of VoiceChannelResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastModifiedDate:VoiceChannelResponse', voiceChannelResponse_lastModifiedDate - The date and time, in ISO 8601 format, when the voice channel was last modified.

$sel:enabled:VoiceChannelResponse', voiceChannelResponse_enabled - Specifies whether the voice channel is enabled for the application.

$sel:isArchived:VoiceChannelResponse', voiceChannelResponse_isArchived - Specifies whether the voice channel is archived.

$sel:applicationId:VoiceChannelResponse', voiceChannelResponse_applicationId - The unique identifier for the application that the voice channel applies to.

$sel:version:VoiceChannelResponse', voiceChannelResponse_version - The current version of the voice channel.

$sel:id:VoiceChannelResponse', voiceChannelResponse_id - (Deprecated) An identifier for the voice channel. This property is retained only for backward compatibility.

$sel:creationDate:VoiceChannelResponse', voiceChannelResponse_creationDate - The date and time, in ISO 8601 format, when the voice channel was enabled.

$sel:lastModifiedBy:VoiceChannelResponse', voiceChannelResponse_lastModifiedBy - The user who last modified the voice channel.

$sel:hasCredential:VoiceChannelResponse', voiceChannelResponse_hasCredential - (Not used) This property is retained only for backward compatibility.

$sel:platform:VoiceChannelResponse', voiceChannelResponse_platform - The type of messaging or notification platform for the channel. For the voice channel, this value is VOICE.

voiceChannelResponse_lastModifiedDate :: Lens' VoiceChannelResponse (Maybe Text)

The date and time, in ISO 8601 format, when the voice channel was last modified.

voiceChannelResponse_enabled :: Lens' VoiceChannelResponse (Maybe Bool)

Specifies whether the voice channel is enabled for the application.

voiceChannelResponse_isArchived :: Lens' VoiceChannelResponse (Maybe Bool)

Specifies whether the voice channel is archived.

voiceChannelResponse_applicationId :: Lens' VoiceChannelResponse (Maybe Text)

The unique identifier for the application that the voice channel applies to.

voiceChannelResponse_version :: Lens' VoiceChannelResponse (Maybe Int)

The current version of the voice channel.

voiceChannelResponse_id :: Lens' VoiceChannelResponse (Maybe Text)

(Deprecated) An identifier for the voice channel. This property is retained only for backward compatibility.

voiceChannelResponse_creationDate :: Lens' VoiceChannelResponse (Maybe Text)

The date and time, in ISO 8601 format, when the voice channel was enabled.

voiceChannelResponse_lastModifiedBy :: Lens' VoiceChannelResponse (Maybe Text)

The user who last modified the voice channel.

voiceChannelResponse_hasCredential :: Lens' VoiceChannelResponse (Maybe Bool)

(Not used) This property is retained only for backward compatibility.

voiceChannelResponse_platform :: Lens' VoiceChannelResponse Text

The type of messaging or notification platform for the channel. For the voice channel, this value is VOICE.