Network.AWS.Pinpoint.Types.VoiceTemplateRequest

Description

Documentation

data VoiceTemplateRequest

Specifies the content and settings for a message template that can be used in messages that are sent through the voice channel.

See: newVoiceTemplateRequest smart constructor.

Constructors

VoiceTemplateRequest'

Fields

languageCode :: Maybe Text

The code for the language to use when synthesizing the text of the script in messages that are based on the message template. For a list of supported languages and the code for each one, see the Amazon Polly Developer Guide.

templateDescription :: Maybe Text

A custom description of the message template.

voiceId :: Maybe Text

The name of the voice to use when delivering messages that are based on the message template. For a list of supported voices, see the Amazon Polly Developer Guide.

body :: Maybe Text

The text of the script to use in messages that are based on the message template, in plain text format.

defaultSubstitutions :: Maybe Text

A JSON object that specifies the default values to use for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable. When you create a message that's based on the template, you can override these defaults with message-specific and address-specific variables and values.

tags :: Maybe (HashMap Text Text)

A string-to-string map of key-value pairs that defines the tags to associate with the message template. Each tag consists of a required tag key and an associated tag value.

newVoiceTemplateRequest :: VoiceTemplateRequest

Create a value of VoiceTemplateRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:VoiceTemplateRequest', voiceTemplateRequest_languageCode - The code for the language to use when synthesizing the text of the script in messages that are based on the message template. For a list of supported languages and the code for each one, see the Amazon Polly Developer Guide.

$sel:templateDescription:VoiceTemplateRequest', voiceTemplateRequest_templateDescription - A custom description of the message template.

$sel:voiceId:VoiceTemplateRequest', voiceTemplateRequest_voiceId - The name of the voice to use when delivering messages that are based on the message template. For a list of supported voices, see the Amazon Polly Developer Guide.

$sel:body:VoiceTemplateRequest', voiceTemplateRequest_body - The text of the script to use in messages that are based on the message template, in plain text format.

$sel:defaultSubstitutions:VoiceTemplateRequest', voiceTemplateRequest_defaultSubstitutions - A JSON object that specifies the default values to use for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable. When you create a message that's based on the template, you can override these defaults with message-specific and address-specific variables and values.

$sel:tags:VoiceTemplateRequest', voiceTemplateRequest_tags - A string-to-string map of key-value pairs that defines the tags to associate with the message template. Each tag consists of a required tag key and an associated tag value.

voiceTemplateRequest_languageCode :: Lens' VoiceTemplateRequest (Maybe Text)

The code for the language to use when synthesizing the text of the script in messages that are based on the message template. For a list of supported languages and the code for each one, see the Amazon Polly Developer Guide.

voiceTemplateRequest_templateDescription :: Lens' VoiceTemplateRequest (Maybe Text)

A custom description of the message template.

voiceTemplateRequest_voiceId :: Lens' VoiceTemplateRequest (Maybe Text)

The name of the voice to use when delivering messages that are based on the message template. For a list of supported voices, see the Amazon Polly Developer Guide.

voiceTemplateRequest_body :: Lens' VoiceTemplateRequest (Maybe Text)

The text of the script to use in messages that are based on the message template, in plain text format.

voiceTemplateRequest_defaultSubstitutions :: Lens' VoiceTemplateRequest (Maybe Text)

A JSON object that specifies the default values to use for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable. When you create a message that's based on the template, you can override these defaults with message-specific and address-specific variables and values.

voiceTemplateRequest_tags :: Lens' VoiceTemplateRequest (Maybe (HashMap Text Text))

A string-to-string map of key-value pairs that defines the tags to associate with the message template. Each tag consists of a required tag key and an associated tag value.