Network.AWS.Pinpoint.Types.VoiceTemplateResponse

Description

Documentation

data VoiceTemplateResponse

Provides information about the content and settings for a message template that can be used in messages that are sent through the voice channel.

See: newVoiceTemplateResponse smart constructor.

Constructors

VoiceTemplateResponse'

Fields

languageCode :: Maybe Text

The code for the language that's used when synthesizing the text of the script in messages that are based on the message template. For a list of supported languages and the code for each one, see the Amazon Polly Developer Guide.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the message template.

body :: Maybe Text

The text of the script that's used in messages that are based on the message template, in plain text format.

templateDescription :: Maybe Text

The custom description of the message template.

defaultSubstitutions :: Maybe Text

The JSON object that specifies the default values that are used for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable.

version :: Maybe Text

The unique identifier, as an integer, for the active version of the message template, or the version of the template that you specified by using the version parameter in your request.

voiceId :: Maybe Text

The name of the voice that's used when delivering messages that are based on the message template. For a list of supported voices, see the Amazon Polly Developer Guide.

tags :: Maybe (HashMap Text Text)

A string-to-string map of key-value pairs that identifies the tags that are associated with the message template. Each tag consists of a required tag key and an associated tag value.

lastModifiedDate :: Text

The date, in ISO 8601 format, when the message template was last modified.

creationDate :: Text

The date, in ISO 8601 format, when the message template was created.

templateName :: Text

The name of the message template.

templateType :: TemplateType

The type of channel that the message template is designed for. For a voice template, this value is VOICE.

newVoiceTemplateResponse

Arguments

:: Text
$sel:lastModifiedDate:VoiceTemplateResponse'

-> Text
$sel:creationDate:VoiceTemplateResponse'

-> Text
$sel:templateName:VoiceTemplateResponse'

-> TemplateType
$sel:templateType:VoiceTemplateResponse'

-> VoiceTemplateResponse

Create a value of VoiceTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:VoiceTemplateResponse', voiceTemplateResponse_languageCode - The code for the language that's used when synthesizing the text of the script in messages that are based on the message template. For a list of supported languages and the code for each one, see the Amazon Polly Developer Guide.

$sel:arn:VoiceTemplateResponse', voiceTemplateResponse_arn - The Amazon Resource Name (ARN) of the message template.

$sel:body:VoiceTemplateResponse', voiceTemplateResponse_body - The text of the script that's used in messages that are based on the message template, in plain text format.

$sel:templateDescription:VoiceTemplateResponse', voiceTemplateResponse_templateDescription - The custom description of the message template.

$sel:defaultSubstitutions:VoiceTemplateResponse', voiceTemplateResponse_defaultSubstitutions - The JSON object that specifies the default values that are used for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable.

$sel:version:VoiceTemplateResponse', voiceTemplateResponse_version - The unique identifier, as an integer, for the active version of the message template, or the version of the template that you specified by using the version parameter in your request.

$sel:voiceId:VoiceTemplateResponse', voiceTemplateResponse_voiceId - The name of the voice that's used when delivering messages that are based on the message template. For a list of supported voices, see the Amazon Polly Developer Guide.

$sel:tags:VoiceTemplateResponse', voiceTemplateResponse_tags - A string-to-string map of key-value pairs that identifies the tags that are associated with the message template. Each tag consists of a required tag key and an associated tag value.

$sel:lastModifiedDate:VoiceTemplateResponse', voiceTemplateResponse_lastModifiedDate - The date, in ISO 8601 format, when the message template was last modified.

$sel:creationDate:VoiceTemplateResponse', voiceTemplateResponse_creationDate - The date, in ISO 8601 format, when the message template was created.

$sel:templateName:VoiceTemplateResponse', voiceTemplateResponse_templateName - The name of the message template.

$sel:templateType:VoiceTemplateResponse', voiceTemplateResponse_templateType - The type of channel that the message template is designed for. For a voice template, this value is VOICE.

voiceTemplateResponse_languageCode :: Lens' VoiceTemplateResponse (Maybe Text)

The code for the language that's used when synthesizing the text of the script in messages that are based on the message template. For a list of supported languages and the code for each one, see the Amazon Polly Developer Guide.

voiceTemplateResponse_arn :: Lens' VoiceTemplateResponse (Maybe Text)

The Amazon Resource Name (ARN) of the message template.

voiceTemplateResponse_body :: Lens' VoiceTemplateResponse (Maybe Text)

The text of the script that's used in messages that are based on the message template, in plain text format.

voiceTemplateResponse_templateDescription :: Lens' VoiceTemplateResponse (Maybe Text)

The custom description of the message template.

voiceTemplateResponse_defaultSubstitutions :: Lens' VoiceTemplateResponse (Maybe Text)

The JSON object that specifies the default values that are used for message variables in the message template. This object is a set of key-value pairs. Each key defines a message variable in the template. The corresponding value defines the default value for that variable.

voiceTemplateResponse_version :: Lens' VoiceTemplateResponse (Maybe Text)

The unique identifier, as an integer, for the active version of the message template, or the version of the template that you specified by using the version parameter in your request.

voiceTemplateResponse_voiceId :: Lens' VoiceTemplateResponse (Maybe Text)

The name of the voice that's used when delivering messages that are based on the message template. For a list of supported voices, see the Amazon Polly Developer Guide.

voiceTemplateResponse_tags :: Lens' VoiceTemplateResponse (Maybe (HashMap Text Text))

A string-to-string map of key-value pairs that identifies the tags that are associated with the message template. Each tag consists of a required tag key and an associated tag value.

voiceTemplateResponse_lastModifiedDate :: Lens' VoiceTemplateResponse Text

The date, in ISO 8601 format, when the message template was last modified.

voiceTemplateResponse_creationDate :: Lens' VoiceTemplateResponse Text

The date, in ISO 8601 format, when the message template was created.

voiceTemplateResponse_templateName :: Lens' VoiceTemplateResponse Text

The name of the message template.

voiceTemplateResponse_templateType :: Lens' VoiceTemplateResponse TemplateType

The type of channel that the message template is designed for. For a voice template, this value is VOICE.