Network.AWS.Pinpoint.Types.WriteApplicationSettingsRequest

Description

Documentation

data WriteApplicationSettingsRequest

Specifies the default settings for an application.

See: newWriteApplicationSettingsRequest smart constructor.

Constructors

WriteApplicationSettingsRequest'

Fields

campaignHook :: Maybe CampaignHook

The settings for the AWS Lambda function to invoke by default as a code hook for campaigns in the application. You can use this hook to customize segments that are used by campaigns in the application.

To override these settings and define custom settings for a specific campaign, use the CampaignHook object of the Campaign resource.

cloudWatchMetricsEnabled :: Maybe Bool

Specifies whether to enable application-related alarms in Amazon CloudWatch.

eventTaggingEnabled :: Maybe Bool

quietTime :: Maybe QuietTime

The default quiet time for campaigns in the application. Quiet time is a specific time range when messages aren't sent to endpoints, if all the following conditions are met:

The EndpointDemographic.Timezone property of the endpoint is set to a valid value.

The current time in the endpoint's time zone is later than or equal to the time specified by the QuietTime.Start property for the application (or a campaign or journey that has custom quiet time settings).

The current time in the endpoint's time zone is earlier than or equal to the time specified by the QuietTime.End property for the application (or a campaign or journey that has custom quiet time settings).

If any of the preceding conditions isn't met, the endpoint will receive messages from a campaign or journey, even if quiet time is enabled.

To override the default quiet time settings for a specific campaign or journey, use the Campaign resource or the Journey resource to define a custom quiet time for the campaign or journey.

limits :: Maybe CampaignLimits

The default sending limits for campaigns in the application. To override these limits and define custom limits for a specific campaign or journey, use the Campaign resource or the Journey resource, respectively.

newWriteApplicationSettingsRequest :: WriteApplicationSettingsRequest

Create a value of WriteApplicationSettingsRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:campaignHook:WriteApplicationSettingsRequest', writeApplicationSettingsRequest_campaignHook - The settings for the AWS Lambda function to invoke by default as a code hook for campaigns in the application. You can use this hook to customize segments that are used by campaigns in the application.

To override these settings and define custom settings for a specific campaign, use the CampaignHook object of the Campaign resource.

$sel:cloudWatchMetricsEnabled:WriteApplicationSettingsRequest', writeApplicationSettingsRequest_cloudWatchMetricsEnabled - Specifies whether to enable application-related alarms in Amazon CloudWatch.

$sel:eventTaggingEnabled:WriteApplicationSettingsRequest', writeApplicationSettingsRequest_eventTaggingEnabled - Undocumented member.

$sel:quietTime:WriteApplicationSettingsRequest', writeApplicationSettingsRequest_quietTime - The default quiet time for campaigns in the application. Quiet time is a specific time range when messages aren't sent to endpoints, if all the following conditions are met:

The EndpointDemographic.Timezone property of the endpoint is set to a valid value.

The current time in the endpoint's time zone is later than or equal to the time specified by the QuietTime.Start property for the application (or a campaign or journey that has custom quiet time settings).

The current time in the endpoint's time zone is earlier than or equal to the time specified by the QuietTime.End property for the application (or a campaign or journey that has custom quiet time settings).

If any of the preceding conditions isn't met, the endpoint will receive messages from a campaign or journey, even if quiet time is enabled.

To override the default quiet time settings for a specific campaign or journey, use the Campaign resource or the Journey resource to define a custom quiet time for the campaign or journey.

$sel:limits:WriteApplicationSettingsRequest', writeApplicationSettingsRequest_limits - The default sending limits for campaigns in the application. To override these limits and define custom limits for a specific campaign or journey, use the Campaign resource or the Journey resource, respectively.

writeApplicationSettingsRequest_campaignHook :: Lens' WriteApplicationSettingsRequest (Maybe CampaignHook)

The settings for the AWS Lambda function to invoke by default as a code hook for campaigns in the application. You can use this hook to customize segments that are used by campaigns in the application.

To override these settings and define custom settings for a specific campaign, use the CampaignHook object of the Campaign resource.

writeApplicationSettingsRequest_cloudWatchMetricsEnabled :: Lens' WriteApplicationSettingsRequest (Maybe Bool)

Specifies whether to enable application-related alarms in Amazon CloudWatch.

writeApplicationSettingsRequest_eventTaggingEnabled :: Lens' WriteApplicationSettingsRequest (Maybe Bool)

Undocumented member.

writeApplicationSettingsRequest_quietTime :: Lens' WriteApplicationSettingsRequest (Maybe QuietTime)

The default quiet time for campaigns in the application. Quiet time is a specific time range when messages aren't sent to endpoints, if all the following conditions are met:

The EndpointDemographic.Timezone property of the endpoint is set to a valid value.

The current time in the endpoint's time zone is later than or equal to the time specified by the QuietTime.Start property for the application (or a campaign or journey that has custom quiet time settings).

The current time in the endpoint's time zone is earlier than or equal to the time specified by the QuietTime.End property for the application (or a campaign or journey that has custom quiet time settings).

If any of the preceding conditions isn't met, the endpoint will receive messages from a campaign or journey, even if quiet time is enabled.

To override the default quiet time settings for a specific campaign or journey, use the Campaign resource or the Journey resource to define a custom quiet time for the campaign or journey.

writeApplicationSettingsRequest_limits :: Lens' WriteApplicationSettingsRequest (Maybe CampaignLimits)

The default sending limits for campaigns in the application. To override these limits and define custom limits for a specific campaign or journey, use the Campaign resource or the Journey resource, respectively.