Network.AWS.Pinpoint.Types.WriteCampaignRequest

Description

Documentation

data WriteCampaignRequest

Specifies the configuration and other settings for a campaign.

See: newWriteCampaignRequest smart constructor.

Constructors

WriteCampaignRequest'

Fields

additionalTreatments :: Maybe [WriteTreatmentResource]

An array of requests that defines additional treatments for the campaign, in addition to the default treatment for the campaign.

hook :: Maybe CampaignHook

The settings for the AWS Lambda function to invoke as a code hook for the campaign. You can use this hook to customize the segment that's used by the campaign.

priority :: Maybe Int

Defines the priority of the campaign, used to decide the order of messages displayed to user if there are multiple messages scheduled to be displayed at the same moment.

name :: Maybe Text

A custom name for the campaign.

customDeliveryConfiguration :: Maybe CustomDeliveryConfiguration

The delivery configuration settings for sending the campaign through a custom channel. This object is required if the MessageConfiguration object for the campaign specifies a CustomMessage object.

isPaused :: Maybe Bool

Specifies whether to pause the campaign. A paused campaign doesn't run unless you resume it by changing this value to false.

segmentVersion :: Maybe Int

The version of the segment to associate with the campaign.

tags :: Maybe (HashMap Text Text)

A string-to-string map of key-value pairs that defines the tags to associate with the campaign. Each tag consists of a required tag key and an associated tag value.

limits :: Maybe CampaignLimits

The messaging limits for the campaign.

description :: Maybe Text

A custom description of the campaign.

segmentId :: Maybe Text

The unique identifier for the segment to associate with the campaign.

messageConfiguration :: Maybe MessageConfiguration

The message configuration settings for the campaign.

treatmentName :: Maybe Text

A custom name of the default treatment for the campaign, if the campaign has multiple treatments. A treatment is a variation of a campaign that's used for A/B testing.

templateConfiguration :: Maybe TemplateConfiguration

The message template to use for the campaign.

schedule :: Maybe Schedule

The schedule settings for the campaign.

holdoutPercent :: Maybe Int

The allocated percentage of users (segment members) who shouldn't receive messages from the campaign.

treatmentDescription :: Maybe Text

A custom description of the default treatment for the campaign.

newWriteCampaignRequest :: WriteCampaignRequest

Create a value of WriteCampaignRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:additionalTreatments:WriteCampaignRequest', writeCampaignRequest_additionalTreatments - An array of requests that defines additional treatments for the campaign, in addition to the default treatment for the campaign.

$sel:hook:WriteCampaignRequest', writeCampaignRequest_hook - The settings for the AWS Lambda function to invoke as a code hook for the campaign. You can use this hook to customize the segment that's used by the campaign.

$sel:priority:WriteCampaignRequest', writeCampaignRequest_priority - Defines the priority of the campaign, used to decide the order of messages displayed to user if there are multiple messages scheduled to be displayed at the same moment.

$sel:name:WriteCampaignRequest', writeCampaignRequest_name - A custom name for the campaign.

$sel:customDeliveryConfiguration:WriteCampaignRequest', writeCampaignRequest_customDeliveryConfiguration - The delivery configuration settings for sending the campaign through a custom channel. This object is required if the MessageConfiguration object for the campaign specifies a CustomMessage object.

$sel:isPaused:WriteCampaignRequest', writeCampaignRequest_isPaused - Specifies whether to pause the campaign. A paused campaign doesn't run unless you resume it by changing this value to false.

$sel:segmentVersion:WriteCampaignRequest', writeCampaignRequest_segmentVersion - The version of the segment to associate with the campaign.

$sel:tags:WriteCampaignRequest', writeCampaignRequest_tags - A string-to-string map of key-value pairs that defines the tags to associate with the campaign. Each tag consists of a required tag key and an associated tag value.

$sel:limits:WriteCampaignRequest', writeCampaignRequest_limits - The messaging limits for the campaign.

$sel:description:WriteCampaignRequest', writeCampaignRequest_description - A custom description of the campaign.

$sel:segmentId:WriteCampaignRequest', writeCampaignRequest_segmentId - The unique identifier for the segment to associate with the campaign.

$sel:messageConfiguration:WriteCampaignRequest', writeCampaignRequest_messageConfiguration - The message configuration settings for the campaign.

$sel:treatmentName:WriteCampaignRequest', writeCampaignRequest_treatmentName - A custom name of the default treatment for the campaign, if the campaign has multiple treatments. A treatment is a variation of a campaign that's used for A/B testing.

$sel:templateConfiguration:WriteCampaignRequest', writeCampaignRequest_templateConfiguration - The message template to use for the campaign.

$sel:schedule:WriteCampaignRequest', writeCampaignRequest_schedule - The schedule settings for the campaign.

$sel:holdoutPercent:WriteCampaignRequest', writeCampaignRequest_holdoutPercent - The allocated percentage of users (segment members) who shouldn't receive messages from the campaign.

$sel:treatmentDescription:WriteCampaignRequest', writeCampaignRequest_treatmentDescription - A custom description of the default treatment for the campaign.

writeCampaignRequest_additionalTreatments :: Lens' WriteCampaignRequest (Maybe [WriteTreatmentResource])

An array of requests that defines additional treatments for the campaign, in addition to the default treatment for the campaign.

writeCampaignRequest_hook :: Lens' WriteCampaignRequest (Maybe CampaignHook)

The settings for the AWS Lambda function to invoke as a code hook for the campaign. You can use this hook to customize the segment that's used by the campaign.

writeCampaignRequest_priority :: Lens' WriteCampaignRequest (Maybe Int)

Defines the priority of the campaign, used to decide the order of messages displayed to user if there are multiple messages scheduled to be displayed at the same moment.

writeCampaignRequest_name :: Lens' WriteCampaignRequest (Maybe Text)

A custom name for the campaign.

writeCampaignRequest_customDeliveryConfiguration :: Lens' WriteCampaignRequest (Maybe CustomDeliveryConfiguration)

The delivery configuration settings for sending the campaign through a custom channel. This object is required if the MessageConfiguration object for the campaign specifies a CustomMessage object.

writeCampaignRequest_isPaused :: Lens' WriteCampaignRequest (Maybe Bool)

Specifies whether to pause the campaign. A paused campaign doesn't run unless you resume it by changing this value to false.

writeCampaignRequest_segmentVersion :: Lens' WriteCampaignRequest (Maybe Int)

The version of the segment to associate with the campaign.

writeCampaignRequest_tags :: Lens' WriteCampaignRequest (Maybe (HashMap Text Text))

A string-to-string map of key-value pairs that defines the tags to associate with the campaign. Each tag consists of a required tag key and an associated tag value.

writeCampaignRequest_limits :: Lens' WriteCampaignRequest (Maybe CampaignLimits)

The messaging limits for the campaign.

writeCampaignRequest_description :: Lens' WriteCampaignRequest (Maybe Text)

A custom description of the campaign.

writeCampaignRequest_segmentId :: Lens' WriteCampaignRequest (Maybe Text)

The unique identifier for the segment to associate with the campaign.

writeCampaignRequest_messageConfiguration :: Lens' WriteCampaignRequest (Maybe MessageConfiguration)

The message configuration settings for the campaign.

writeCampaignRequest_treatmentName :: Lens' WriteCampaignRequest (Maybe Text)

A custom name of the default treatment for the campaign, if the campaign has multiple treatments. A treatment is a variation of a campaign that's used for A/B testing.

writeCampaignRequest_templateConfiguration :: Lens' WriteCampaignRequest (Maybe TemplateConfiguration)

The message template to use for the campaign.

writeCampaignRequest_schedule :: Lens' WriteCampaignRequest (Maybe Schedule)

The schedule settings for the campaign.

writeCampaignRequest_holdoutPercent :: Lens' WriteCampaignRequest (Maybe Int)

The allocated percentage of users (segment members) who shouldn't receive messages from the campaign.

writeCampaignRequest_treatmentDescription :: Lens' WriteCampaignRequest (Maybe Text)

A custom description of the default treatment for the campaign.