Network.AWS.Pinpoint.Types.WriteJourneyRequest

Description

Documentation

data WriteJourneyRequest

Specifies the configuration and other settings for a journey.

See: newWriteJourneyRequest smart constructor.

Constructors

WriteJourneyRequest'

Fields

state :: Maybe State

The status of the journey. Valid values are:

DRAFT - Saves the journey and doesn't publish it.

ACTIVE - Saves and publishes the journey. Depending on the journey's schedule, the journey starts running immediately or at the scheduled start time. If a journey's status is ACTIVE, you can't add, change, or remove activities from it.

PAUSED, CANCELLED, COMPLETED, and CLOSED states are not supported in requests to create or update a journey. To cancel, pause, or resume a journey, use the Journey State resource.

lastModifiedDate :: Maybe Text

The date, in ISO 8601 format, when the journey was last modified.

schedule :: Maybe JourneySchedule

The schedule settings for the journey.

localTime :: Maybe Bool

Specifies whether the journey's scheduled start and end times use each participant's local time. To base the schedule on each participant's local time, set this value to true.

activities :: Maybe (HashMap Text Activity)

A map that contains a set of Activity objects, one object for each activity in the journey. For each Activity object, the key is the unique identifier (string) for an activity and the value is the settings for the activity. An activity identifier can contain a maximum of 100 characters. The characters must be alphanumeric characters.

refreshOnSegmentUpdate :: Maybe Bool

Specifies whether a journey should be refreshed on segment update.

limits :: Maybe JourneyLimits

The messaging and entry limits for the journey.

waitForQuietTime :: Maybe Bool

Specifies whether endpoints in quiet hours should enter a wait till the end of their quiet hours.

quietTime :: Maybe QuietTime

The quiet time settings for the journey. Quiet time is a specific time range when a journey doesn't send messages to participants, if all the following conditions are met:

The EndpointDemographic.Timezone property of the endpoint for the participant is set to a valid value.

The current time in the participant's time zone is later than or equal to the time specified by the QuietTime.Start property for the journey.

The current time in the participant's time zone is earlier than or equal to the time specified by the QuietTime.End property for the journey.

If any of the preceding conditions isn't met, the participant will receive messages from the journey, even if quiet time is enabled.

startActivity :: Maybe Text

The unique identifier for the first activity in the journey. The identifier for this activity can contain a maximum of 128 characters. The characters must be alphanumeric characters.

creationDate :: Maybe Text

The date, in ISO 8601 format, when the journey was created.

startCondition :: Maybe StartCondition

The segment that defines which users are participants in the journey.

refreshFrequency :: Maybe Text

The frequency with which Amazon Pinpoint evaluates segment and event data for the journey, as a duration in ISO 8601 format.

name :: Text

The name of the journey. A journey name can contain a maximum of 150 characters. The characters can be alphanumeric characters or symbols, such as underscores (_) or hyphens (-). A journey name can't contain any spaces.

newWriteJourneyRequest

Arguments

:: Text
$sel:name:WriteJourneyRequest'

-> WriteJourneyRequest

Create a value of WriteJourneyRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:WriteJourneyRequest', writeJourneyRequest_state - The status of the journey. Valid values are:

DRAFT - Saves the journey and doesn't publish it.

ACTIVE - Saves and publishes the journey. Depending on the journey's schedule, the journey starts running immediately or at the scheduled start time. If a journey's status is ACTIVE, you can't add, change, or remove activities from it.

PAUSED, CANCELLED, COMPLETED, and CLOSED states are not supported in requests to create or update a journey. To cancel, pause, or resume a journey, use the Journey State resource.

$sel:lastModifiedDate:WriteJourneyRequest', writeJourneyRequest_lastModifiedDate - The date, in ISO 8601 format, when the journey was last modified.

$sel:schedule:WriteJourneyRequest', writeJourneyRequest_schedule - The schedule settings for the journey.

$sel:localTime:WriteJourneyRequest', writeJourneyRequest_localTime - Specifies whether the journey's scheduled start and end times use each participant's local time. To base the schedule on each participant's local time, set this value to true.

$sel:activities:WriteJourneyRequest', writeJourneyRequest_activities - A map that contains a set of Activity objects, one object for each activity in the journey. For each Activity object, the key is the unique identifier (string) for an activity and the value is the settings for the activity. An activity identifier can contain a maximum of 100 characters. The characters must be alphanumeric characters.

$sel:refreshOnSegmentUpdate:WriteJourneyRequest', writeJourneyRequest_refreshOnSegmentUpdate - Specifies whether a journey should be refreshed on segment update.

$sel:limits:WriteJourneyRequest', writeJourneyRequest_limits - The messaging and entry limits for the journey.

$sel:waitForQuietTime:WriteJourneyRequest', writeJourneyRequest_waitForQuietTime - Specifies whether endpoints in quiet hours should enter a wait till the end of their quiet hours.

$sel:quietTime:WriteJourneyRequest', writeJourneyRequest_quietTime - The quiet time settings for the journey. Quiet time is a specific time range when a journey doesn't send messages to participants, if all the following conditions are met:

The EndpointDemographic.Timezone property of the endpoint for the participant is set to a valid value.

The current time in the participant's time zone is later than or equal to the time specified by the QuietTime.Start property for the journey.

The current time in the participant's time zone is earlier than or equal to the time specified by the QuietTime.End property for the journey.

If any of the preceding conditions isn't met, the participant will receive messages from the journey, even if quiet time is enabled.

$sel:startActivity:WriteJourneyRequest', writeJourneyRequest_startActivity - The unique identifier for the first activity in the journey. The identifier for this activity can contain a maximum of 128 characters. The characters must be alphanumeric characters.

$sel:creationDate:WriteJourneyRequest', writeJourneyRequest_creationDate - The date, in ISO 8601 format, when the journey was created.

$sel:startCondition:WriteJourneyRequest', writeJourneyRequest_startCondition - The segment that defines which users are participants in the journey.

$sel:refreshFrequency:WriteJourneyRequest', writeJourneyRequest_refreshFrequency - The frequency with which Amazon Pinpoint evaluates segment and event data for the journey, as a duration in ISO 8601 format.

$sel:name:WriteJourneyRequest', writeJourneyRequest_name - The name of the journey. A journey name can contain a maximum of 150 characters. The characters can be alphanumeric characters or symbols, such as underscores (_) or hyphens (-). A journey name can't contain any spaces.

writeJourneyRequest_state :: Lens' WriteJourneyRequest (Maybe State)

The status of the journey. Valid values are:

DRAFT - Saves the journey and doesn't publish it.

ACTIVE - Saves and publishes the journey. Depending on the journey's schedule, the journey starts running immediately or at the scheduled start time. If a journey's status is ACTIVE, you can't add, change, or remove activities from it.

PAUSED, CANCELLED, COMPLETED, and CLOSED states are not supported in requests to create or update a journey. To cancel, pause, or resume a journey, use the Journey State resource.

writeJourneyRequest_lastModifiedDate :: Lens' WriteJourneyRequest (Maybe Text)

The date, in ISO 8601 format, when the journey was last modified.

writeJourneyRequest_schedule :: Lens' WriteJourneyRequest (Maybe JourneySchedule)

The schedule settings for the journey.

writeJourneyRequest_localTime :: Lens' WriteJourneyRequest (Maybe Bool)

Specifies whether the journey's scheduled start and end times use each participant's local time. To base the schedule on each participant's local time, set this value to true.

writeJourneyRequest_activities :: Lens' WriteJourneyRequest (Maybe (HashMap Text Activity))

A map that contains a set of Activity objects, one object for each activity in the journey. For each Activity object, the key is the unique identifier (string) for an activity and the value is the settings for the activity. An activity identifier can contain a maximum of 100 characters. The characters must be alphanumeric characters.

writeJourneyRequest_refreshOnSegmentUpdate :: Lens' WriteJourneyRequest (Maybe Bool)

Specifies whether a journey should be refreshed on segment update.

writeJourneyRequest_limits :: Lens' WriteJourneyRequest (Maybe JourneyLimits)

The messaging and entry limits for the journey.

writeJourneyRequest_waitForQuietTime :: Lens' WriteJourneyRequest (Maybe Bool)

Specifies whether endpoints in quiet hours should enter a wait till the end of their quiet hours.

writeJourneyRequest_quietTime :: Lens' WriteJourneyRequest (Maybe QuietTime)

The quiet time settings for the journey. Quiet time is a specific time range when a journey doesn't send messages to participants, if all the following conditions are met:

The EndpointDemographic.Timezone property of the endpoint for the participant is set to a valid value.

The current time in the participant's time zone is later than or equal to the time specified by the QuietTime.Start property for the journey.

The current time in the participant's time zone is earlier than or equal to the time specified by the QuietTime.End property for the journey.

If any of the preceding conditions isn't met, the participant will receive messages from the journey, even if quiet time is enabled.

writeJourneyRequest_startActivity :: Lens' WriteJourneyRequest (Maybe Text)

The unique identifier for the first activity in the journey. The identifier for this activity can contain a maximum of 128 characters. The characters must be alphanumeric characters.

writeJourneyRequest_creationDate :: Lens' WriteJourneyRequest (Maybe Text)

The date, in ISO 8601 format, when the journey was created.

writeJourneyRequest_startCondition :: Lens' WriteJourneyRequest (Maybe StartCondition)

The segment that defines which users are participants in the journey.

writeJourneyRequest_refreshFrequency :: Lens' WriteJourneyRequest (Maybe Text)

The frequency with which Amazon Pinpoint evaluates segment and event data for the journey, as a duration in ISO 8601 format.

writeJourneyRequest_name :: Lens' WriteJourneyRequest Text

The name of the journey. A journey name can contain a maximum of 150 characters. The characters can be alphanumeric characters or symbols, such as underscores (_) or hyphens (-). A journey name can't contain any spaces.