Network.AWS.Pinpoint.Types.WriteTreatmentResource

Description

Documentation

data WriteTreatmentResource

Specifies the settings for a campaign treatment. A treatment is a variation of a campaign that's used for A/B testing of a campaign.

See: newWriteTreatmentResource smart constructor.

Constructors

WriteTreatmentResource'

Fields

customDeliveryConfiguration :: Maybe CustomDeliveryConfiguration

The delivery configuration settings for sending the treatment through a custom channel. This object is required if the MessageConfiguration object for the treatment specifies a CustomMessage object.

messageConfiguration :: Maybe MessageConfiguration

The message configuration settings for the treatment.

treatmentName :: Maybe Text

A custom name for the treatment.

templateConfiguration :: Maybe TemplateConfiguration

The message template to use for the treatment.

schedule :: Maybe Schedule

The schedule settings for the treatment.

treatmentDescription :: Maybe Text

A custom description of the treatment.

sizePercent :: Int

The allocated percentage of users (segment members) to send the treatment to.

newWriteTreatmentResource

Arguments

:: Int
$sel:sizePercent:WriteTreatmentResource'

-> WriteTreatmentResource

Create a value of WriteTreatmentResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customDeliveryConfiguration:WriteTreatmentResource', writeTreatmentResource_customDeliveryConfiguration - The delivery configuration settings for sending the treatment through a custom channel. This object is required if the MessageConfiguration object for the treatment specifies a CustomMessage object.

$sel:messageConfiguration:WriteTreatmentResource', writeTreatmentResource_messageConfiguration - The message configuration settings for the treatment.

$sel:treatmentName:WriteTreatmentResource', writeTreatmentResource_treatmentName - A custom name for the treatment.

$sel:templateConfiguration:WriteTreatmentResource', writeTreatmentResource_templateConfiguration - The message template to use for the treatment.

$sel:schedule:WriteTreatmentResource', writeTreatmentResource_schedule - The schedule settings for the treatment.

$sel:treatmentDescription:WriteTreatmentResource', writeTreatmentResource_treatmentDescription - A custom description of the treatment.

$sel:sizePercent:WriteTreatmentResource', writeTreatmentResource_sizePercent - The allocated percentage of users (segment members) to send the treatment to.

writeTreatmentResource_customDeliveryConfiguration :: Lens' WriteTreatmentResource (Maybe CustomDeliveryConfiguration)

The delivery configuration settings for sending the treatment through a custom channel. This object is required if the MessageConfiguration object for the treatment specifies a CustomMessage object.

writeTreatmentResource_messageConfiguration :: Lens' WriteTreatmentResource (Maybe MessageConfiguration)

The message configuration settings for the treatment.

writeTreatmentResource_treatmentName :: Lens' WriteTreatmentResource (Maybe Text)

A custom name for the treatment.

writeTreatmentResource_templateConfiguration :: Lens' WriteTreatmentResource (Maybe TemplateConfiguration)

The message template to use for the treatment.

writeTreatmentResource_schedule :: Lens' WriteTreatmentResource (Maybe Schedule)

The schedule settings for the treatment.

writeTreatmentResource_treatmentDescription :: Lens' WriteTreatmentResource (Maybe Text)

A custom description of the treatment.

writeTreatmentResource_sizePercent :: Lens' WriteTreatmentResource Int

The allocated percentage of users (segment members) to send the treatment to.