Network.AWS.PinpointEmail.CreateConfigurationSetEventDestination

Description

Create an event destination. In Amazon Pinpoint, events include message sends, deliveries, opens, clicks, bounces, and complaints. Event destinations are places that you can send information about these events to. For example, you can send event data to Amazon SNS to receive notifications when you receive bounces or complaints, or you can use Amazon Kinesis Data Firehose to stream data to Amazon S3 for long-term storage.

A single configuration set can include more than one event destination.

Creating a Request

data CreateConfigurationSetEventDestination

A request to add an event destination to a configuration set.

See: newCreateConfigurationSetEventDestination smart constructor.

Constructors

CreateConfigurationSetEventDestination'

Fields

configurationSetName :: Text

The name of the configuration set that you want to add an event destination to.

eventDestinationName :: Text

A name that identifies the event destination within the configuration set.

eventDestination :: EventDestinationDefinition

An object that defines the event destination.

newCreateConfigurationSetEventDestination

Arguments

:: Text
$sel:configurationSetName:CreateConfigurationSetEventDestination'

-> Text
$sel:eventDestinationName:CreateConfigurationSetEventDestination'

-> EventDestinationDefinition
$sel:eventDestination:CreateConfigurationSetEventDestination'

-> CreateConfigurationSetEventDestination

Create a value of CreateConfigurationSetEventDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationSetName:CreateConfigurationSetEventDestination', createConfigurationSetEventDestination_configurationSetName - The name of the configuration set that you want to add an event destination to.

$sel:eventDestinationName:CreateConfigurationSetEventDestination', createConfigurationSetEventDestination_eventDestinationName - A name that identifies the event destination within the configuration set.

$sel:eventDestination:CreateConfigurationSetEventDestination', createConfigurationSetEventDestination_eventDestination - An object that defines the event destination.

Request Lenses

createConfigurationSetEventDestination_configurationSetName :: Lens' CreateConfigurationSetEventDestination Text

The name of the configuration set that you want to add an event destination to.

createConfigurationSetEventDestination_eventDestinationName :: Lens' CreateConfigurationSetEventDestination Text

A name that identifies the event destination within the configuration set.

createConfigurationSetEventDestination_eventDestination :: Lens' CreateConfigurationSetEventDestination EventDestinationDefinition

An object that defines the event destination.

Destructuring the Response

data CreateConfigurationSetEventDestinationResponse

An HTTP 200 response if the request succeeds, or an error message if the request fails.

See: newCreateConfigurationSetEventDestinationResponse smart constructor.

Constructors

CreateConfigurationSetEventDestinationResponse'

Fields

httpStatus :: Int

The response's http status code.

newCreateConfigurationSetEventDestinationResponse

Arguments

:: Int
$sel:httpStatus:CreateConfigurationSetEventDestinationResponse'

-> CreateConfigurationSetEventDestinationResponse

Create a value of CreateConfigurationSetEventDestinationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateConfigurationSetEventDestinationResponse', createConfigurationSetEventDestinationResponse_httpStatus - The response's http status code.

Response Lenses

createConfigurationSetEventDestinationResponse_httpStatus :: Lens' CreateConfigurationSetEventDestinationResponse Int

The response's http status code.