Network.AWS.PinpointEmail.CreateDeliverabilityTestReport

Description

Create a new predictive inbox placement test. Predictive inbox placement tests can help you predict how your messages will be handled by various email providers around the world. When you perform a predictive inbox placement test, you provide a sample message that contains the content that you plan to send to your customers. Amazon Pinpoint then sends that message to special email addresses spread across several major email providers. After about 24 hours, the test is complete, and you can use the GetDeliverabilityTestReport operation to view the results of the test.

Creating a Request

data CreateDeliverabilityTestReport

A request to perform a predictive inbox placement test. Predictive inbox placement tests can help you predict how your messages will be handled by various email providers around the world. When you perform a predictive inbox placement test, you provide a sample message that contains the content that you plan to send to your customers. Amazon Pinpoint then sends that message to special email addresses spread across several major email providers. After about 24 hours, the test is complete, and you can use the GetDeliverabilityTestReport operation to view the results of the test.

See: newCreateDeliverabilityTestReport smart constructor.

Constructors

CreateDeliverabilityTestReport'

Fields

reportName :: Maybe Text

A unique name that helps you to identify the predictive inbox placement test when you retrieve the results.

tags :: Maybe [Tag]

An array of objects that define the tags (keys and values) that you want to associate with the predictive inbox placement test.

fromEmailAddress :: Text

The email address that the predictive inbox placement test email was sent from.

content :: EmailContent

The HTML body of the message that you sent when you performed the predictive inbox placement test.

newCreateDeliverabilityTestReport

Arguments

:: Text
$sel:fromEmailAddress:CreateDeliverabilityTestReport'

-> EmailContent
$sel:content:CreateDeliverabilityTestReport'

-> CreateDeliverabilityTestReport

Create a value of CreateDeliverabilityTestReport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reportName:CreateDeliverabilityTestReport', createDeliverabilityTestReport_reportName - A unique name that helps you to identify the predictive inbox placement test when you retrieve the results.

$sel:tags:CreateDeliverabilityTestReport', createDeliverabilityTestReport_tags - An array of objects that define the tags (keys and values) that you want to associate with the predictive inbox placement test.

$sel:fromEmailAddress:CreateDeliverabilityTestReport', createDeliverabilityTestReport_fromEmailAddress - The email address that the predictive inbox placement test email was sent from.

$sel:content:CreateDeliverabilityTestReport', createDeliverabilityTestReport_content - The HTML body of the message that you sent when you performed the predictive inbox placement test.

Request Lenses

createDeliverabilityTestReport_reportName :: Lens' CreateDeliverabilityTestReport (Maybe Text)

A unique name that helps you to identify the predictive inbox placement test when you retrieve the results.

createDeliverabilityTestReport_tags :: Lens' CreateDeliverabilityTestReport (Maybe [Tag])

An array of objects that define the tags (keys and values) that you want to associate with the predictive inbox placement test.

createDeliverabilityTestReport_fromEmailAddress :: Lens' CreateDeliverabilityTestReport Text

The email address that the predictive inbox placement test email was sent from.

createDeliverabilityTestReport_content :: Lens' CreateDeliverabilityTestReport EmailContent

The HTML body of the message that you sent when you performed the predictive inbox placement test.

Destructuring the Response

data CreateDeliverabilityTestReportResponse

Information about the predictive inbox placement test that you created.

See: newCreateDeliverabilityTestReportResponse smart constructor.

Constructors

CreateDeliverabilityTestReportResponse'

Fields

httpStatus :: Int

The response's http status code.

reportId :: Text

A unique string that identifies the predictive inbox placement test.

deliverabilityTestStatus :: DeliverabilityTestStatus

The status of the predictive inbox placement test. If the status is IN_PROGRESS, then the predictive inbox placement test is currently running. Predictive inbox placement tests are usually complete within 24 hours of creating the test. If the status is COMPLETE, then the test is finished, and you can use the GetDeliverabilityTestReport to view the results of the test.

newCreateDeliverabilityTestReportResponse

Arguments

:: Int
$sel:httpStatus:CreateDeliverabilityTestReportResponse'

-> Text
$sel:reportId:CreateDeliverabilityTestReportResponse'

-> DeliverabilityTestStatus
$sel:deliverabilityTestStatus:CreateDeliverabilityTestReportResponse'

-> CreateDeliverabilityTestReportResponse

Create a value of CreateDeliverabilityTestReportResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:CreateDeliverabilityTestReportResponse', createDeliverabilityTestReportResponse_httpStatus - The response's http status code.

$sel:reportId:CreateDeliverabilityTestReportResponse', createDeliverabilityTestReportResponse_reportId - A unique string that identifies the predictive inbox placement test.

$sel:deliverabilityTestStatus:CreateDeliverabilityTestReportResponse', createDeliverabilityTestReportResponse_deliverabilityTestStatus - The status of the predictive inbox placement test. If the status is IN_PROGRESS, then the predictive inbox placement test is currently running. Predictive inbox placement tests are usually complete within 24 hours of creating the test. If the status is COMPLETE, then the test is finished, and you can use the GetDeliverabilityTestReport to view the results of the test.

Response Lenses

createDeliverabilityTestReportResponse_httpStatus :: Lens' CreateDeliverabilityTestReportResponse Int

The response's http status code.

createDeliverabilityTestReportResponse_reportId :: Lens' CreateDeliverabilityTestReportResponse Text

A unique string that identifies the predictive inbox placement test.

createDeliverabilityTestReportResponse_deliverabilityTestStatus :: Lens' CreateDeliverabilityTestReportResponse DeliverabilityTestStatus

The status of the predictive inbox placement test. If the status is IN_PROGRESS, then the predictive inbox placement test is currently running. Predictive inbox placement tests are usually complete within 24 hours of creating the test. If the status is COMPLETE, then the test is finished, and you can use the GetDeliverabilityTestReport to view the results of the test.