Network.AWS.PinpointEmail.PutConfigurationSetSendingOptions

Description

Enable or disable email sending for messages that use a particular configuration set in a specific AWS Region.

Creating a Request

data PutConfigurationSetSendingOptions

A request to enable or disable the ability of Amazon Pinpoint to send emails that use a specific configuration set.

See: newPutConfigurationSetSendingOptions smart constructor.

Constructors

PutConfigurationSetSendingOptions'

Fields

sendingEnabled :: Maybe Bool

If true, email sending is enabled for the configuration set. If false, email sending is disabled for the configuration set.

configurationSetName :: Text

The name of the configuration set that you want to enable or disable email sending for.

newPutConfigurationSetSendingOptions

Arguments

:: Text
$sel:configurationSetName:PutConfigurationSetSendingOptions'

-> PutConfigurationSetSendingOptions

Create a value of PutConfigurationSetSendingOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sendingEnabled:PutConfigurationSetSendingOptions', putConfigurationSetSendingOptions_sendingEnabled - If true, email sending is enabled for the configuration set. If false, email sending is disabled for the configuration set.

$sel:configurationSetName:PutConfigurationSetSendingOptions', putConfigurationSetSendingOptions_configurationSetName - The name of the configuration set that you want to enable or disable email sending for.

Request Lenses

putConfigurationSetSendingOptions_sendingEnabled :: Lens' PutConfigurationSetSendingOptions (Maybe Bool)

If true, email sending is enabled for the configuration set. If false, email sending is disabled for the configuration set.

putConfigurationSetSendingOptions_configurationSetName :: Lens' PutConfigurationSetSendingOptions Text

The name of the configuration set that you want to enable or disable email sending for.

Destructuring the Response

data PutConfigurationSetSendingOptionsResponse

An HTTP 200 response if the request succeeds, or an error message if the request fails.

See: newPutConfigurationSetSendingOptionsResponse smart constructor.

Constructors

PutConfigurationSetSendingOptionsResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutConfigurationSetSendingOptionsResponse

Arguments

:: Int
$sel:httpStatus:PutConfigurationSetSendingOptionsResponse'

-> PutConfigurationSetSendingOptionsResponse

Create a value of PutConfigurationSetSendingOptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutConfigurationSetSendingOptionsResponse', putConfigurationSetSendingOptionsResponse_httpStatus - The response's http status code.

Response Lenses

putConfigurationSetSendingOptionsResponse_httpStatus :: Lens' PutConfigurationSetSendingOptionsResponse Int

The response's http status code.