Network.AWS.PinpointEmail.Types.CloudWatchDimensionConfiguration

Description

Documentation

data CloudWatchDimensionConfiguration

An object that defines the dimension configuration to use when you send Amazon Pinpoint email events to Amazon CloudWatch.

See: newCloudWatchDimensionConfiguration smart constructor.

Constructors

CloudWatchDimensionConfiguration'

Fields

dimensionName :: Text

The name of an Amazon CloudWatch dimension associated with an email sending metric. The name has to meet the following criteria:

It can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

It can contain no more than 256 characters.

dimensionValueSource :: DimensionValueSource

The location where Amazon Pinpoint finds the value of a dimension to publish to Amazon CloudWatch. If you want Amazon Pinpoint to use the message tags that you specify using an X-SES-MESSAGE-TAGS header or a parameter to the SendEmail/SendRawEmail API, choose messageTag. If you want Amazon Pinpoint to use your own email headers, choose emailHeader. If you want Amazon Pinpoint to use link tags, choose linkTags.

defaultDimensionValue :: Text

The default value of the dimension that is published to Amazon CloudWatch if you don't provide the value of the dimension when you send an email. This value has to meet the following criteria:

It can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

It can contain no more than 256 characters.

newCloudWatchDimensionConfiguration

Arguments

:: Text
$sel:dimensionName:CloudWatchDimensionConfiguration'

-> DimensionValueSource
$sel:dimensionValueSource:CloudWatchDimensionConfiguration'

-> Text
$sel:defaultDimensionValue:CloudWatchDimensionConfiguration'

-> CloudWatchDimensionConfiguration

Create a value of CloudWatchDimensionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dimensionName:CloudWatchDimensionConfiguration', cloudWatchDimensionConfiguration_dimensionName - The name of an Amazon CloudWatch dimension associated with an email sending metric. The name has to meet the following criteria:

It can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

It can contain no more than 256 characters.

$sel:dimensionValueSource:CloudWatchDimensionConfiguration', cloudWatchDimensionConfiguration_dimensionValueSource - The location where Amazon Pinpoint finds the value of a dimension to publish to Amazon CloudWatch. If you want Amazon Pinpoint to use the message tags that you specify using an X-SES-MESSAGE-TAGS header or a parameter to the SendEmail/SendRawEmail API, choose messageTag. If you want Amazon Pinpoint to use your own email headers, choose emailHeader. If you want Amazon Pinpoint to use link tags, choose linkTags.

$sel:defaultDimensionValue:CloudWatchDimensionConfiguration', cloudWatchDimensionConfiguration_defaultDimensionValue - The default value of the dimension that is published to Amazon CloudWatch if you don't provide the value of the dimension when you send an email. This value has to meet the following criteria:

It can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

It can contain no more than 256 characters.

cloudWatchDimensionConfiguration_dimensionName :: Lens' CloudWatchDimensionConfiguration Text

The name of an Amazon CloudWatch dimension associated with an email sending metric. The name has to meet the following criteria:

It can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

It can contain no more than 256 characters.

cloudWatchDimensionConfiguration_dimensionValueSource :: Lens' CloudWatchDimensionConfiguration DimensionValueSource

The location where Amazon Pinpoint finds the value of a dimension to publish to Amazon CloudWatch. If you want Amazon Pinpoint to use the message tags that you specify using an X-SES-MESSAGE-TAGS header or a parameter to the SendEmail/SendRawEmail API, choose messageTag. If you want Amazon Pinpoint to use your own email headers, choose emailHeader. If you want Amazon Pinpoint to use link tags, choose linkTags.

cloudWatchDimensionConfiguration_defaultDimensionValue :: Lens' CloudWatchDimensionConfiguration Text

The default value of the dimension that is published to Amazon CloudWatch if you don't provide the value of the dimension when you send an email. This value has to meet the following criteria:

It can only contain ASCII letters (a-z, A-Z), numbers (0-9), underscores (_), or dashes (-).

It can contain no more than 256 characters.