Network.AWS.PinpointEmail.Types.Content

Description

Documentation

data Content

An object that represents the content of the email, and optionally a character set specification.

See: newContent smart constructor.

Constructors

Content'

Fields

charset :: Maybe Text

The character set for the content. Because of the constraints of the SMTP protocol, Amazon Pinpoint uses 7-bit ASCII by default. If the text includes characters outside of the ASCII range, you have to specify a character set. For example, you could specify UTF-8, ISO-8859-1, or Shift_JIS.

data' :: Text

The content of the message itself.

newContent

Arguments

:: Text
$sel:data':Content'

-> Content

Create a value of Content with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:charset:Content', content_charset - The character set for the content. Because of the constraints of the SMTP protocol, Amazon Pinpoint uses 7-bit ASCII by default. If the text includes characters outside of the ASCII range, you have to specify a character set. For example, you could specify UTF-8, ISO-8859-1, or Shift_JIS.

$sel:data':Content', content_data - The content of the message itself.

content_charset :: Lens' Content (Maybe Text)

The character set for the content. Because of the constraints of the SMTP protocol, Amazon Pinpoint uses 7-bit ASCII by default. If the text includes characters outside of the ASCII range, you have to specify a character set. For example, you could specify UTF-8, ISO-8859-1, or Shift_JIS.

content_data :: Lens' Content Text

The content of the message itself.