Network.AWS.PinpointEmail.Types.DeliverabilityTestReport

Description

Documentation

data DeliverabilityTestReport

An object that contains metadata related to a predictive inbox placement test.

See: newDeliverabilityTestReport smart constructor.

Constructors

DeliverabilityTestReport'

Fields

subject :: Maybe Text

The subject line for an email that you submitted in a predictive inbox placement test.

fromEmailAddress :: Maybe Text

The sender address that you specified for the predictive inbox placement test.

createDate :: Maybe POSIX

The date and time when the predictive inbox placement test was created, in Unix time format.

reportId :: Maybe Text

A unique string that identifies the predictive inbox placement test.

reportName :: Maybe Text

A name that helps you identify a predictive inbox placement test report.

deliverabilityTestStatus :: Maybe DeliverabilityTestStatus

The status of the predictive inbox placement test. If the status is IN_PROGRESS, then the predictive inbox placement test is currently running. Predictive inbox placement tests are usually complete within 24 hours of creating the test. If the status is COMPLETE, then the test is finished, and you can use the GetDeliverabilityTestReport to view the results of the test.

newDeliverabilityTestReport :: DeliverabilityTestReport

Create a value of DeliverabilityTestReport with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subject:DeliverabilityTestReport', deliverabilityTestReport_subject - The subject line for an email that you submitted in a predictive inbox placement test.

$sel:fromEmailAddress:DeliverabilityTestReport', deliverabilityTestReport_fromEmailAddress - The sender address that you specified for the predictive inbox placement test.

$sel:createDate:DeliverabilityTestReport', deliverabilityTestReport_createDate - The date and time when the predictive inbox placement test was created, in Unix time format.

$sel:reportId:DeliverabilityTestReport', deliverabilityTestReport_reportId - A unique string that identifies the predictive inbox placement test.

$sel:reportName:DeliverabilityTestReport', deliverabilityTestReport_reportName - A name that helps you identify a predictive inbox placement test report.

$sel:deliverabilityTestStatus:DeliverabilityTestReport', deliverabilityTestReport_deliverabilityTestStatus - The status of the predictive inbox placement test. If the status is IN_PROGRESS, then the predictive inbox placement test is currently running. Predictive inbox placement tests are usually complete within 24 hours of creating the test. If the status is COMPLETE, then the test is finished, and you can use the GetDeliverabilityTestReport to view the results of the test.

deliverabilityTestReport_subject :: Lens' DeliverabilityTestReport (Maybe Text)

The subject line for an email that you submitted in a predictive inbox placement test.

deliverabilityTestReport_fromEmailAddress :: Lens' DeliverabilityTestReport (Maybe Text)

The sender address that you specified for the predictive inbox placement test.

deliverabilityTestReport_createDate :: Lens' DeliverabilityTestReport (Maybe UTCTime)

The date and time when the predictive inbox placement test was created, in Unix time format.

deliverabilityTestReport_reportId :: Lens' DeliverabilityTestReport (Maybe Text)

A unique string that identifies the predictive inbox placement test.

deliverabilityTestReport_reportName :: Lens' DeliverabilityTestReport (Maybe Text)

A name that helps you identify a predictive inbox placement test report.

deliverabilityTestReport_deliverabilityTestStatus :: Lens' DeliverabilityTestReport (Maybe DeliverabilityTestStatus)

The status of the predictive inbox placement test. If the status is IN_PROGRESS, then the predictive inbox placement test is currently running. Predictive inbox placement tests are usually complete within 24 hours of creating the test. If the status is COMPLETE, then the test is finished, and you can use the GetDeliverabilityTestReport to view the results of the test.