Network.AWS.PinpointEmail.Types.DomainDeliverabilityCampaign

Description

Documentation

data DomainDeliverabilityCampaign

An object that contains the deliverability data for a specific campaign. This data is available for a campaign only if the campaign sent email by using a domain that the Deliverability dashboard is enabled for (PutDeliverabilityDashboardOption operation).

See: newDomainDeliverabilityCampaign smart constructor.

Constructors

DomainDeliverabilityCampaign'

Fields

spamCount :: Maybe Integer

The number of email messages that were delivered to recipients' spam or junk mail folders.

subject :: Maybe Text

The subject line, or title, of the email message.

esps :: Maybe [Text]

The major email providers who handled the email message.

fromAddress :: Maybe Text

The verified email address that the email message was sent from.

deleteRate :: Maybe Double

The percentage of email messages that were deleted by recipients, without being opened first. Due to technical limitations, this value only includes recipients who opened the message by using an email client that supports images.

campaignId :: Maybe Text

The unique identifier for the campaign. Amazon Pinpoint automatically generates and assigns this identifier to a campaign. This value is not the same as the campaign identifier that Amazon Pinpoint assigns to campaigns that you create and manage by using the Amazon Pinpoint API or the Amazon Pinpoint console.

sendingIps :: Maybe [Text]

The IP addresses that were used to send the email message.

firstSeenDateTime :: Maybe POSIX

The first time, in Unix time format, when the email message was delivered to any recipient's inbox. This value can help you determine how long it took for a campaign to deliver an email message.

inboxCount :: Maybe Integer

The number of email messages that were delivered to recipients’ inboxes.

readDeleteRate :: Maybe Double

The percentage of email messages that were opened and then deleted by recipients. Due to technical limitations, this value only includes recipients who opened the message by using an email client that supports images.

projectedVolume :: Maybe Integer

The projected number of recipients that the email message was sent to.

imageUrl :: Maybe Text

The URL of an image that contains a snapshot of the email message that was sent.

readRate :: Maybe Double

The percentage of email messages that were opened by recipients. Due to technical limitations, this value only includes recipients who opened the message by using an email client that supports images.

lastSeenDateTime :: Maybe POSIX

The last time, in Unix time format, when the email message was delivered to any recipient's inbox. This value can help you determine how long it took for a campaign to deliver an email message.

newDomainDeliverabilityCampaign :: DomainDeliverabilityCampaign

Create a value of DomainDeliverabilityCampaign with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:spamCount:DomainDeliverabilityCampaign', domainDeliverabilityCampaign_spamCount - The number of email messages that were delivered to recipients' spam or junk mail folders.

$sel:subject:DomainDeliverabilityCampaign', domainDeliverabilityCampaign_subject - The subject line, or title, of the email message.

$sel:esps:DomainDeliverabilityCampaign', domainDeliverabilityCampaign_esps - The major email providers who handled the email message.

$sel:fromAddress:DomainDeliverabilityCampaign', domainDeliverabilityCampaign_fromAddress - The verified email address that the email message was sent from.

$sel:deleteRate:DomainDeliverabilityCampaign', domainDeliverabilityCampaign_deleteRate - The percentage of email messages that were deleted by recipients, without being opened first. Due to technical limitations, this value only includes recipients who opened the message by using an email client that supports images.

$sel:campaignId:DomainDeliverabilityCampaign', domainDeliverabilityCampaign_campaignId - The unique identifier for the campaign. Amazon Pinpoint automatically generates and assigns this identifier to a campaign. This value is not the same as the campaign identifier that Amazon Pinpoint assigns to campaigns that you create and manage by using the Amazon Pinpoint API or the Amazon Pinpoint console.

$sel:sendingIps:DomainDeliverabilityCampaign', domainDeliverabilityCampaign_sendingIps - The IP addresses that were used to send the email message.

$sel:firstSeenDateTime:DomainDeliverabilityCampaign', domainDeliverabilityCampaign_firstSeenDateTime - The first time, in Unix time format, when the email message was delivered to any recipient's inbox. This value can help you determine how long it took for a campaign to deliver an email message.

$sel:inboxCount:DomainDeliverabilityCampaign', domainDeliverabilityCampaign_inboxCount - The number of email messages that were delivered to recipients’ inboxes.

$sel:readDeleteRate:DomainDeliverabilityCampaign', domainDeliverabilityCampaign_readDeleteRate - The percentage of email messages that were opened and then deleted by recipients. Due to technical limitations, this value only includes recipients who opened the message by using an email client that supports images.

$sel:projectedVolume:DomainDeliverabilityCampaign', domainDeliverabilityCampaign_projectedVolume - The projected number of recipients that the email message was sent to.

$sel:imageUrl:DomainDeliverabilityCampaign', domainDeliverabilityCampaign_imageUrl - The URL of an image that contains a snapshot of the email message that was sent.

$sel:readRate:DomainDeliverabilityCampaign', domainDeliverabilityCampaign_readRate - The percentage of email messages that were opened by recipients. Due to technical limitations, this value only includes recipients who opened the message by using an email client that supports images.

$sel:lastSeenDateTime:DomainDeliverabilityCampaign', domainDeliverabilityCampaign_lastSeenDateTime - The last time, in Unix time format, when the email message was delivered to any recipient's inbox. This value can help you determine how long it took for a campaign to deliver an email message.

domainDeliverabilityCampaign_spamCount :: Lens' DomainDeliverabilityCampaign (Maybe Integer)

The number of email messages that were delivered to recipients' spam or junk mail folders.

domainDeliverabilityCampaign_subject :: Lens' DomainDeliverabilityCampaign (Maybe Text)

The subject line, or title, of the email message.

domainDeliverabilityCampaign_esps :: Lens' DomainDeliverabilityCampaign (Maybe [Text])

The major email providers who handled the email message.

domainDeliverabilityCampaign_fromAddress :: Lens' DomainDeliverabilityCampaign (Maybe Text)

The verified email address that the email message was sent from.

domainDeliverabilityCampaign_deleteRate :: Lens' DomainDeliverabilityCampaign (Maybe Double)

The percentage of email messages that were deleted by recipients, without being opened first. Due to technical limitations, this value only includes recipients who opened the message by using an email client that supports images.

domainDeliverabilityCampaign_campaignId :: Lens' DomainDeliverabilityCampaign (Maybe Text)

The unique identifier for the campaign. Amazon Pinpoint automatically generates and assigns this identifier to a campaign. This value is not the same as the campaign identifier that Amazon Pinpoint assigns to campaigns that you create and manage by using the Amazon Pinpoint API or the Amazon Pinpoint console.

domainDeliverabilityCampaign_sendingIps :: Lens' DomainDeliverabilityCampaign (Maybe [Text])

The IP addresses that were used to send the email message.

domainDeliverabilityCampaign_firstSeenDateTime :: Lens' DomainDeliverabilityCampaign (Maybe UTCTime)

The first time, in Unix time format, when the email message was delivered to any recipient's inbox. This value can help you determine how long it took for a campaign to deliver an email message.

domainDeliverabilityCampaign_inboxCount :: Lens' DomainDeliverabilityCampaign (Maybe Integer)

The number of email messages that were delivered to recipients’ inboxes.

domainDeliverabilityCampaign_readDeleteRate :: Lens' DomainDeliverabilityCampaign (Maybe Double)

The percentage of email messages that were opened and then deleted by recipients. Due to technical limitations, this value only includes recipients who opened the message by using an email client that supports images.

domainDeliverabilityCampaign_projectedVolume :: Lens' DomainDeliverabilityCampaign (Maybe Integer)

The projected number of recipients that the email message was sent to.

domainDeliverabilityCampaign_imageUrl :: Lens' DomainDeliverabilityCampaign (Maybe Text)

The URL of an image that contains a snapshot of the email message that was sent.

domainDeliverabilityCampaign_readRate :: Lens' DomainDeliverabilityCampaign (Maybe Double)

The percentage of email messages that were opened by recipients. Due to technical limitations, this value only includes recipients who opened the message by using an email client that supports images.

domainDeliverabilityCampaign_lastSeenDateTime :: Lens' DomainDeliverabilityCampaign (Maybe UTCTime)

The last time, in Unix time format, when the email message was delivered to any recipient's inbox. This value can help you determine how long it took for a campaign to deliver an email message.