Network.AWS.PinpointEmail.Types.MailFromAttributes

Description

Documentation

data MailFromAttributes

A list of attributes that are associated with a MAIL FROM domain.

See: newMailFromAttributes smart constructor.

Constructors

MailFromAttributes'

Fields

mailFromDomain :: Text

The name of a domain that an email identity uses as a custom MAIL FROM domain.

mailFromDomainStatus :: MailFromDomainStatus

The status of the MAIL FROM domain. This status can have the following values:

PENDING – Amazon Pinpoint hasn't started searching for the MX record yet.

SUCCESS – Amazon Pinpoint detected the required MX record for the MAIL FROM domain.

FAILED – Amazon Pinpoint can't find the required MX record, or the record no longer exists.

TEMPORARY_FAILURE – A temporary issue occurred, which prevented Amazon Pinpoint from determining the status of the MAIL FROM domain.

behaviorOnMxFailure :: BehaviorOnMxFailure

The action that Amazon Pinpoint to takes if it can't read the required MX record for a custom MAIL FROM domain. When you set this value to UseDefaultValue, Amazon Pinpoint uses amazonses.com as the MAIL FROM domain. When you set this value to RejectMessage, Amazon Pinpoint returns a MailFromDomainNotVerified error, and doesn't attempt to deliver the email.

These behaviors are taken when the custom MAIL FROM domain configuration is in the Pending, Failed, and TemporaryFailure states.

newMailFromAttributes

Arguments

:: Text
$sel:mailFromDomain:MailFromAttributes'

-> MailFromDomainStatus
$sel:mailFromDomainStatus:MailFromAttributes'

-> BehaviorOnMxFailure
$sel:behaviorOnMxFailure:MailFromAttributes'

-> MailFromAttributes

Create a value of MailFromAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:mailFromDomain:MailFromAttributes', mailFromAttributes_mailFromDomain - The name of a domain that an email identity uses as a custom MAIL FROM domain.

$sel:mailFromDomainStatus:MailFromAttributes', mailFromAttributes_mailFromDomainStatus - The status of the MAIL FROM domain. This status can have the following values:

PENDING – Amazon Pinpoint hasn't started searching for the MX record yet.

SUCCESS – Amazon Pinpoint detected the required MX record for the MAIL FROM domain.

FAILED – Amazon Pinpoint can't find the required MX record, or the record no longer exists.

TEMPORARY_FAILURE – A temporary issue occurred, which prevented Amazon Pinpoint from determining the status of the MAIL FROM domain.

$sel:behaviorOnMxFailure:MailFromAttributes', mailFromAttributes_behaviorOnMxFailure - The action that Amazon Pinpoint to takes if it can't read the required MX record for a custom MAIL FROM domain. When you set this value to UseDefaultValue, Amazon Pinpoint uses amazonses.com as the MAIL FROM domain. When you set this value to RejectMessage, Amazon Pinpoint returns a MailFromDomainNotVerified error, and doesn't attempt to deliver the email.

These behaviors are taken when the custom MAIL FROM domain configuration is in the Pending, Failed, and TemporaryFailure states.

mailFromAttributes_mailFromDomain :: Lens' MailFromAttributes Text

The name of a domain that an email identity uses as a custom MAIL FROM domain.

mailFromAttributes_mailFromDomainStatus :: Lens' MailFromAttributes MailFromDomainStatus

The status of the MAIL FROM domain. This status can have the following values:

PENDING – Amazon Pinpoint hasn't started searching for the MX record yet.

SUCCESS – Amazon Pinpoint detected the required MX record for the MAIL FROM domain.

FAILED – Amazon Pinpoint can't find the required MX record, or the record no longer exists.

TEMPORARY_FAILURE – A temporary issue occurred, which prevented Amazon Pinpoint from determining the status of the MAIL FROM domain.

mailFromAttributes_behaviorOnMxFailure :: Lens' MailFromAttributes BehaviorOnMxFailure

The action that Amazon Pinpoint to takes if it can't read the required MX record for a custom MAIL FROM domain. When you set this value to UseDefaultValue, Amazon Pinpoint uses amazonses.com as the MAIL FROM domain. When you set this value to RejectMessage, Amazon Pinpoint returns a MailFromDomainNotVerified error, and doesn't attempt to deliver the email.

These behaviors are taken when the custom MAIL FROM domain configuration is in the Pending, Failed, and TemporaryFailure states.