Network.AWS.PinpointEmail.UntagResource

Description

Remove one or more tags (keys and values) from a specified resource.

Creating a Request

data UntagResource

See: newUntagResource smart constructor.

Constructors

UntagResource'

Fields

resourceArn :: Text

The Amazon Resource Name (ARN) of the resource that you want to remove one or more tags from.

tagKeys :: [Text]

The tags (tag keys) that you want to remove from the resource. When you specify a tag key, the action removes both that key and its associated tag value.

To remove more than one tag from the resource, append the TagKeys parameter and argument for each additional tag to remove, separated by an ampersand. For example: /v1/email/tags?ResourceArn=ResourceArn&TagKeys=Key1&TagKeys=Key2

newUntagResource

Arguments

:: Text
$sel:resourceArn:UntagResource'

-> UntagResource

Create a value of UntagResource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:UntagResource', untagResource_resourceArn - The Amazon Resource Name (ARN) of the resource that you want to remove one or more tags from.

$sel:tagKeys:UntagResource', untagResource_tagKeys - The tags (tag keys) that you want to remove from the resource. When you specify a tag key, the action removes both that key and its associated tag value.

To remove more than one tag from the resource, append the TagKeys parameter and argument for each additional tag to remove, separated by an ampersand. For example: /v1/email/tags?ResourceArn=ResourceArn&TagKeys=Key1&TagKeys=Key2

Request Lenses

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) of the resource that you want to remove one or more tags from.

untagResource_tagKeys :: Lens' UntagResource [Text]

The tags (tag keys) that you want to remove from the resource. When you specify a tag key, the action removes both that key and its associated tag value.

To remove more than one tag from the resource, append the TagKeys parameter and argument for each additional tag to remove, separated by an ampersand. For example: /v1/email/tags?ResourceArn=ResourceArn&TagKeys=Key1&TagKeys=Key2

Destructuring the Response

data UntagResourceResponse

See: newUntagResourceResponse smart constructor.

Constructors

UntagResourceResponse'

Fields

httpStatus :: Int

The response's http status code.

newUntagResourceResponse

Arguments

:: Int
$sel:httpStatus:UntagResourceResponse'

-> UntagResourceResponse

Create a value of UntagResourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UntagResourceResponse', untagResourceResponse_httpStatus - The response's http status code.

Response Lenses

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.