Network.AWS.PinpointSMSVoice.DeleteConfigurationSetEventDestination

Description

Deletes an event destination in a configuration set.

Creating a Request

data DeleteConfigurationSetEventDestination

See: newDeleteConfigurationSetEventDestination smart constructor.

Constructors

DeleteConfigurationSetEventDestination'

Fields

eventDestinationName :: Text

EventDestinationName

configurationSetName :: Text

ConfigurationSetName

newDeleteConfigurationSetEventDestination

Arguments

:: Text
$sel:eventDestinationName:DeleteConfigurationSetEventDestination'

-> Text
$sel:configurationSetName:DeleteConfigurationSetEventDestination'

-> DeleteConfigurationSetEventDestination

Create a value of DeleteConfigurationSetEventDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventDestinationName:DeleteConfigurationSetEventDestination', deleteConfigurationSetEventDestination_eventDestinationName - EventDestinationName

$sel:configurationSetName:DeleteConfigurationSetEventDestination', deleteConfigurationSetEventDestination_configurationSetName - ConfigurationSetName

Request Lenses

deleteConfigurationSetEventDestination_eventDestinationName :: Lens' DeleteConfigurationSetEventDestination Text

EventDestinationName

deleteConfigurationSetEventDestination_configurationSetName :: Lens' DeleteConfigurationSetEventDestination Text

ConfigurationSetName

Destructuring the Response

data DeleteConfigurationSetEventDestinationResponse

An empty object that indicates that the event destination was deleted successfully.

See: newDeleteConfigurationSetEventDestinationResponse smart constructor.

Constructors

DeleteConfigurationSetEventDestinationResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeleteConfigurationSetEventDestinationResponse

Arguments

:: Int
$sel:httpStatus:DeleteConfigurationSetEventDestinationResponse'

-> DeleteConfigurationSetEventDestinationResponse

Create a value of DeleteConfigurationSetEventDestinationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeleteConfigurationSetEventDestinationResponse', deleteConfigurationSetEventDestinationResponse_httpStatus - The response's http status code.

Response Lenses

deleteConfigurationSetEventDestinationResponse_httpStatus :: Lens' DeleteConfigurationSetEventDestinationResponse Int

The response's http status code.