Network.AWS.PinpointSMSVoice.SendVoiceMessage

Description

Create a new voice message and send it to a recipient's phone number.

Creating a Request

data SendVoiceMessage

SendVoiceMessageRequest

See: newSendVoiceMessage smart constructor.

Constructors

SendVoiceMessage'

Fields

configurationSetName :: Maybe Text

The name of the configuration set that you want to use to send the message.

callerId :: Maybe Text

The phone number that appears on recipients' devices when they receive the message.

originationPhoneNumber :: Maybe Text

The phone number that Amazon Pinpoint should use to send the voice message. This isn't necessarily the phone number that appears on recipients' devices when they receive the message, because you can specify a CallerId parameter in the request.

content :: Maybe VoiceMessageContent

destinationPhoneNumber :: Maybe Text

The phone number that you want to send the voice message to.

newSendVoiceMessage :: SendVoiceMessage

Create a value of SendVoiceMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:configurationSetName:SendVoiceMessage', sendVoiceMessage_configurationSetName - The name of the configuration set that you want to use to send the message.

$sel:callerId:SendVoiceMessage', sendVoiceMessage_callerId - The phone number that appears on recipients' devices when they receive the message.

$sel:originationPhoneNumber:SendVoiceMessage', sendVoiceMessage_originationPhoneNumber - The phone number that Amazon Pinpoint should use to send the voice message. This isn't necessarily the phone number that appears on recipients' devices when they receive the message, because you can specify a CallerId parameter in the request.

$sel:content:SendVoiceMessage', sendVoiceMessage_content - Undocumented member.

$sel:destinationPhoneNumber:SendVoiceMessage', sendVoiceMessage_destinationPhoneNumber - The phone number that you want to send the voice message to.

Request Lenses

sendVoiceMessage_configurationSetName :: Lens' SendVoiceMessage (Maybe Text)

The name of the configuration set that you want to use to send the message.

sendVoiceMessage_callerId :: Lens' SendVoiceMessage (Maybe Text)

The phone number that appears on recipients' devices when they receive the message.

sendVoiceMessage_originationPhoneNumber :: Lens' SendVoiceMessage (Maybe Text)

The phone number that Amazon Pinpoint should use to send the voice message. This isn't necessarily the phone number that appears on recipients' devices when they receive the message, because you can specify a CallerId parameter in the request.

sendVoiceMessage_content :: Lens' SendVoiceMessage (Maybe VoiceMessageContent)

Undocumented member.

sendVoiceMessage_destinationPhoneNumber :: Lens' SendVoiceMessage (Maybe Text)

The phone number that you want to send the voice message to.

Destructuring the Response

data SendVoiceMessageResponse

An object that that contains the Message ID of a Voice message that was sent successfully.

See: newSendVoiceMessageResponse smart constructor.

Constructors

SendVoiceMessageResponse'

Fields

messageId :: Maybe Text

A unique identifier for the voice message.

httpStatus :: Int

The response's http status code.

newSendVoiceMessageResponse

Arguments

:: Int
$sel:httpStatus:SendVoiceMessageResponse'

-> SendVoiceMessageResponse

Create a value of SendVoiceMessageResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:messageId:SendVoiceMessageResponse', sendVoiceMessageResponse_messageId - A unique identifier for the voice message.

$sel:httpStatus:SendVoiceMessageResponse', sendVoiceMessageResponse_httpStatus - The response's http status code.

Response Lenses

sendVoiceMessageResponse_messageId :: Lens' SendVoiceMessageResponse (Maybe Text)

A unique identifier for the voice message.

sendVoiceMessageResponse_httpStatus :: Lens' SendVoiceMessageResponse Int

The response's http status code.