Network.AWS.PinpointSMSVoice.UpdateConfigurationSetEventDestination

Description

Update an event destination in a configuration set. An event destination is a location that you publish information about your voice calls to. For example, you can log an event to an Amazon CloudWatch destination when a call fails.

Creating a Request

data UpdateConfigurationSetEventDestination

UpdateConfigurationSetEventDestinationRequest

See: newUpdateConfigurationSetEventDestination smart constructor.

Constructors

UpdateConfigurationSetEventDestination'

Fields

eventDestination :: Maybe EventDestinationDefinition

eventDestinationName :: Text

EventDestinationName

configurationSetName :: Text

ConfigurationSetName

newUpdateConfigurationSetEventDestination

Arguments

:: Text
$sel:eventDestinationName:UpdateConfigurationSetEventDestination'

-> Text
$sel:configurationSetName:UpdateConfigurationSetEventDestination'

-> UpdateConfigurationSetEventDestination

Create a value of UpdateConfigurationSetEventDestination with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventDestination:UpdateConfigurationSetEventDestination', updateConfigurationSetEventDestination_eventDestination - Undocumented member.

$sel:eventDestinationName:UpdateConfigurationSetEventDestination', updateConfigurationSetEventDestination_eventDestinationName - EventDestinationName

$sel:configurationSetName:UpdateConfigurationSetEventDestination', updateConfigurationSetEventDestination_configurationSetName - ConfigurationSetName

Request Lenses

updateConfigurationSetEventDestination_eventDestination :: Lens' UpdateConfigurationSetEventDestination (Maybe EventDestinationDefinition)

Undocumented member.

updateConfigurationSetEventDestination_eventDestinationName :: Lens' UpdateConfigurationSetEventDestination Text

EventDestinationName

updateConfigurationSetEventDestination_configurationSetName :: Lens' UpdateConfigurationSetEventDestination Text

ConfigurationSetName

Destructuring the Response

data UpdateConfigurationSetEventDestinationResponse

An empty object that indicates that the event destination was updated successfully.

See: newUpdateConfigurationSetEventDestinationResponse smart constructor.

Constructors

UpdateConfigurationSetEventDestinationResponse'

Fields

httpStatus :: Int

The response's http status code.

newUpdateConfigurationSetEventDestinationResponse

Arguments

:: Int
$sel:httpStatus:UpdateConfigurationSetEventDestinationResponse'

-> UpdateConfigurationSetEventDestinationResponse

Create a value of UpdateConfigurationSetEventDestinationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateConfigurationSetEventDestinationResponse', updateConfigurationSetEventDestinationResponse_httpStatus - The response's http status code.

Response Lenses

updateConfigurationSetEventDestinationResponse_httpStatus :: Lens' UpdateConfigurationSetEventDestinationResponse Int

The response's http status code.