Network.AWS.Polly.GetSpeechSynthesisTask

Description

Retrieves a specific SpeechSynthesisTask object based on its TaskID. This object contains information about the given speech synthesis task, including the status of the task, and a link to the S3 bucket containing the output of the task.

Creating a Request

data GetSpeechSynthesisTask

See: newGetSpeechSynthesisTask smart constructor.

Constructors

GetSpeechSynthesisTask'

Fields

taskId :: Text

The Amazon Polly generated identifier for a speech synthesis task.

newGetSpeechSynthesisTask

Arguments

:: Text
$sel:taskId:GetSpeechSynthesisTask'

-> GetSpeechSynthesisTask

Create a value of GetSpeechSynthesisTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskId:GetSpeechSynthesisTask', getSpeechSynthesisTask_taskId - The Amazon Polly generated identifier for a speech synthesis task.

Request Lenses

getSpeechSynthesisTask_taskId :: Lens' GetSpeechSynthesisTask Text

The Amazon Polly generated identifier for a speech synthesis task.

Destructuring the Response

data GetSpeechSynthesisTaskResponse

See: newGetSpeechSynthesisTaskResponse smart constructor.

Constructors

GetSpeechSynthesisTaskResponse'

Fields

synthesisTask :: Maybe SynthesisTask

SynthesisTask object that provides information from the requested task, including output format, creation time, task status, and so on.

httpStatus :: Int

The response's http status code.

newGetSpeechSynthesisTaskResponse

Arguments

:: Int
$sel:httpStatus:GetSpeechSynthesisTaskResponse'

-> GetSpeechSynthesisTaskResponse

Create a value of GetSpeechSynthesisTaskResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:synthesisTask:GetSpeechSynthesisTaskResponse', getSpeechSynthesisTaskResponse_synthesisTask - SynthesisTask object that provides information from the requested task, including output format, creation time, task status, and so on.

$sel:httpStatus:GetSpeechSynthesisTaskResponse', getSpeechSynthesisTaskResponse_httpStatus - The response's http status code.

Response Lenses

getSpeechSynthesisTaskResponse_synthesisTask :: Lens' GetSpeechSynthesisTaskResponse (Maybe SynthesisTask)

SynthesisTask object that provides information from the requested task, including output format, creation time, task status, and so on.

getSpeechSynthesisTaskResponse_httpStatus :: Lens' GetSpeechSynthesisTaskResponse Int

The response's http status code.