Network.AWS.Polly.Lens

Description

Operations

GetSpeechSynthesisTask

getSpeechSynthesisTask_taskId :: Lens' GetSpeechSynthesisTask Text

The Amazon Polly generated identifier for a speech synthesis task.

getSpeechSynthesisTaskResponse_synthesisTask :: Lens' GetSpeechSynthesisTaskResponse (Maybe SynthesisTask)

SynthesisTask object that provides information from the requested task, including output format, creation time, task status, and so on.

getSpeechSynthesisTaskResponse_httpStatus :: Lens' GetSpeechSynthesisTaskResponse Int

The response's http status code.

GetLexicon

getLexicon_name :: Lens' GetLexicon Text

Name of the lexicon.

getLexiconResponse_lexiconAttributes :: Lens' GetLexiconResponse (Maybe LexiconAttributes)

Metadata of the lexicon, including phonetic alphabetic used, language code, lexicon ARN, number of lexemes defined in the lexicon, and size of lexicon in bytes.

getLexiconResponse_lexicon :: Lens' GetLexiconResponse (Maybe Lexicon)

Lexicon object that provides name and the string content of the lexicon.

getLexiconResponse_httpStatus :: Lens' GetLexiconResponse Int

The response's http status code.

DeleteLexicon

deleteLexicon_name :: Lens' DeleteLexicon Text

The name of the lexicon to delete. Must be an existing lexicon in the region.

deleteLexiconResponse_httpStatus :: Lens' DeleteLexiconResponse Int

The response's http status code.

ListLexicons

listLexicons_nextToken :: Lens' ListLexicons (Maybe Text)

An opaque pagination token returned from previous ListLexicons operation. If present, indicates where to continue the list of lexicons.

listLexiconsResponse_nextToken :: Lens' ListLexiconsResponse (Maybe Text)

The pagination token to use in the next request to continue the listing of lexicons. NextToken is returned only if the response is truncated.

listLexiconsResponse_lexicons :: Lens' ListLexiconsResponse (Maybe [LexiconDescription])

A list of lexicon names and attributes.

listLexiconsResponse_httpStatus :: Lens' ListLexiconsResponse Int

The response's http status code.

DescribeVoices

describeVoices_languageCode :: Lens' DescribeVoices (Maybe LanguageCode)

The language identification tag (ISO 639 code for the language name-ISO 3166 country code) for filtering the list of voices returned. If you don't specify this optional parameter, all available voices are returned.

describeVoices_nextToken :: Lens' DescribeVoices (Maybe Text)

An opaque pagination token returned from the previous DescribeVoices operation. If present, this indicates where to continue the listing.

describeVoices_includeAdditionalLanguageCodes :: Lens' DescribeVoices (Maybe Bool)

Boolean value indicating whether to return any bilingual voices that use the specified language as an additional language. For instance, if you request all languages that use US English (es-US), and there is an Italian voice that speaks both Italian (it-IT) and US English, that voice will be included if you specify yes but not if you specify no.

describeVoices_engine :: Lens' DescribeVoices (Maybe Engine)

Specifies the engine (standard or neural) used by Amazon Polly when processing input text for speech synthesis.

describeVoicesResponse_nextToken :: Lens' DescribeVoicesResponse (Maybe Text)

The pagination token to use in the next request to continue the listing of voices. NextToken is returned only if the response is truncated.

describeVoicesResponse_voices :: Lens' DescribeVoicesResponse (Maybe [Voice])

A list of voices with their properties.

describeVoicesResponse_httpStatus :: Lens' DescribeVoicesResponse Int

The response's http status code.

StartSpeechSynthesisTask

startSpeechSynthesisTask_languageCode :: Lens' StartSpeechSynthesisTask (Maybe LanguageCode)

Optional language code for the Speech Synthesis request. This is only necessary if using a bilingual voice, such as Aditi, which can be used for either Indian English (en-IN) or Hindi (hi-IN).

If a bilingual voice is used and no language code is specified, Amazon Polly uses the default language of the bilingual voice. The default language for any voice is the one returned by the DescribeVoices operation for the LanguageCode parameter. For example, if no language code is specified, Aditi will use Indian English rather than Hindi.

startSpeechSynthesisTask_lexiconNames :: Lens' StartSpeechSynthesisTask (Maybe [Text])

List of one or more pronunciation lexicon names you want the service to apply during synthesis. Lexicons are applied only if the language of the lexicon is the same as the language of the voice.

startSpeechSynthesisTask_speechMarkTypes :: Lens' StartSpeechSynthesisTask (Maybe [SpeechMarkType])

The type of speech marks returned for the input text.

startSpeechSynthesisTask_textType :: Lens' StartSpeechSynthesisTask (Maybe TextType)

Specifies whether the input text is plain text or SSML. The default value is plain text.

startSpeechSynthesisTask_sampleRate :: Lens' StartSpeechSynthesisTask (Maybe Text)

The audio frequency specified in Hz.

The valid values for mp3 and ogg_vorbis are "8000", "16000", "22050", and "24000". The default value for standard voices is "22050". The default value for neural voices is "24000".

Valid values for pcm are "8000" and "16000" The default value is "16000".

startSpeechSynthesisTask_engine :: Lens' StartSpeechSynthesisTask (Maybe Engine)

Specifies the engine (standard or neural) for Amazon Polly to use when processing input text for speech synthesis. Using a voice that is not supported for the engine selected will result in an error.

startSpeechSynthesisTask_outputS3KeyPrefix :: Lens' StartSpeechSynthesisTask (Maybe Text)

The Amazon S3 key prefix for the output speech file.

startSpeechSynthesisTask_snsTopicArn :: Lens' StartSpeechSynthesisTask (Maybe Text)

ARN for the SNS topic optionally used for providing status notification for a speech synthesis task.

startSpeechSynthesisTask_outputFormat :: Lens' StartSpeechSynthesisTask OutputFormat

The format in which the returned output will be encoded. For audio stream, this will be mp3, ogg_vorbis, or pcm. For speech marks, this will be json.

startSpeechSynthesisTask_outputS3BucketName :: Lens' StartSpeechSynthesisTask Text

Amazon S3 bucket name to which the output file will be saved.

startSpeechSynthesisTask_text :: Lens' StartSpeechSynthesisTask Text

The input text to synthesize. If you specify ssml as the TextType, follow the SSML format for the input text.

startSpeechSynthesisTask_voiceId :: Lens' StartSpeechSynthesisTask VoiceId

Voice ID to use for the synthesis.

startSpeechSynthesisTaskResponse_synthesisTask :: Lens' StartSpeechSynthesisTaskResponse (Maybe SynthesisTask)

SynthesisTask object that provides information and attributes about a newly submitted speech synthesis task.

startSpeechSynthesisTaskResponse_httpStatus :: Lens' StartSpeechSynthesisTaskResponse Int

The response's http status code.

PutLexicon

putLexicon_name :: Lens' PutLexicon Text

Name of the lexicon. The name must follow the regular express format [0-9A-Za-z]{1,20}. That is, the name is a case-sensitive alphanumeric string up to 20 characters long.

putLexicon_content :: Lens' PutLexicon Text

Content of the PLS lexicon as string data.

putLexiconResponse_httpStatus :: Lens' PutLexiconResponse Int

The response's http status code.

SynthesizeSpeech

synthesizeSpeech_languageCode :: Lens' SynthesizeSpeech (Maybe LanguageCode)

Optional language code for the Synthesize Speech request. This is only necessary if using a bilingual voice, such as Aditi, which can be used for either Indian English (en-IN) or Hindi (hi-IN).

If a bilingual voice is used and no language code is specified, Amazon Polly uses the default language of the bilingual voice. The default language for any voice is the one returned by the DescribeVoices operation for the LanguageCode parameter. For example, if no language code is specified, Aditi will use Indian English rather than Hindi.

synthesizeSpeech_lexiconNames :: Lens' SynthesizeSpeech (Maybe [Text])

List of one or more pronunciation lexicon names you want the service to apply during synthesis. Lexicons are applied only if the language of the lexicon is the same as the language of the voice. For information about storing lexicons, see PutLexicon.

synthesizeSpeech_speechMarkTypes :: Lens' SynthesizeSpeech (Maybe [SpeechMarkType])

The type of speech marks returned for the input text.

synthesizeSpeech_textType :: Lens' SynthesizeSpeech (Maybe TextType)

Specifies whether the input text is plain text or SSML. The default value is plain text. For more information, see Using SSML.

synthesizeSpeech_sampleRate :: Lens' SynthesizeSpeech (Maybe Text)

The audio frequency specified in Hz.

The valid values for mp3 and ogg_vorbis are "8000", "16000", "22050", and "24000". The default value for standard voices is "22050". The default value for neural voices is "24000".

Valid values for pcm are "8000" and "16000" The default value is "16000".

synthesizeSpeech_engine :: Lens' SynthesizeSpeech (Maybe Engine)

Specifies the engine (standard or neural) for Amazon Polly to use when processing input text for speech synthesis. For information on Amazon Polly voices and which voices are available in standard-only, NTTS-only, and both standard and NTTS formats, see Available Voices.

NTTS-only voices

When using NTTS-only voices such as Kevin (en-US), this parameter is required and must be set to neural. If the engine is not specified, or is set to standard, this will result in an error.

Type: String

Valid Values: standard | neural

Required: Yes

Standard voices

For standard voices, this is not required; the engine parameter defaults to standard. If the engine is not specified, or is set to standard and an NTTS-only voice is selected, this will result in an error.

synthesizeSpeech_outputFormat :: Lens' SynthesizeSpeech OutputFormat

The format in which the returned output will be encoded. For audio stream, this will be mp3, ogg_vorbis, or pcm. For speech marks, this will be json.

When pcm is used, the content returned is audio/pcm in a signed 16-bit, 1 channel (mono), little-endian format.

synthesizeSpeech_text :: Lens' SynthesizeSpeech Text

Input text to synthesize. If you specify ssml as the TextType, follow the SSML format for the input text.

synthesizeSpeech_voiceId :: Lens' SynthesizeSpeech VoiceId

Voice ID to use for the synthesis. You can get a list of available voice IDs by calling the DescribeVoices operation.

synthesizeSpeechResponse_contentType :: Lens' SynthesizeSpeechResponse (Maybe Text)

Specifies the type audio stream. This should reflect the OutputFormat parameter in your request.

If you request mp3 as the OutputFormat, the ContentType returned is audio/mpeg.

If you request ogg_vorbis as the OutputFormat, the ContentType returned is audio/ogg.

If you request pcm as the OutputFormat, the ContentType returned is audio/pcm in a signed 16-bit, 1 channel (mono), little-endian format.

If you request json as the OutputFormat, the ContentType returned is audio/json.

synthesizeSpeechResponse_requestCharacters :: Lens' SynthesizeSpeechResponse (Maybe Int)

Number of characters synthesized.

synthesizeSpeechResponse_httpStatus :: Lens' SynthesizeSpeechResponse Int

The response's http status code.

synthesizeSpeechResponse_audioStream :: Lens' SynthesizeSpeechResponse ResponseBody

Stream containing the synthesized speech.

ListSpeechSynthesisTasks

listSpeechSynthesisTasks_status :: Lens' ListSpeechSynthesisTasks (Maybe TaskStatus)

Status of the speech synthesis tasks returned in a List operation

listSpeechSynthesisTasks_nextToken :: Lens' ListSpeechSynthesisTasks (Maybe Text)

The pagination token to use in the next request to continue the listing of speech synthesis tasks.

listSpeechSynthesisTasks_maxResults :: Lens' ListSpeechSynthesisTasks (Maybe Natural)

Maximum number of speech synthesis tasks returned in a List operation.

listSpeechSynthesisTasksResponse_nextToken :: Lens' ListSpeechSynthesisTasksResponse (Maybe Text)

An opaque pagination token returned from the previous List operation in this request. If present, this indicates where to continue the listing.

listSpeechSynthesisTasksResponse_synthesisTasks :: Lens' ListSpeechSynthesisTasksResponse (Maybe [SynthesisTask])

List of SynthesisTask objects that provides information from the specified task in the list request, including output format, creation time, task status, and so on.

listSpeechSynthesisTasksResponse_httpStatus :: Lens' ListSpeechSynthesisTasksResponse Int

The response's http status code.

Types

Lexicon

lexicon_name :: Lens' Lexicon (Maybe Text)

Name of the lexicon.

lexicon_content :: Lens' Lexicon (Maybe Text)

Lexicon content in string format. The content of a lexicon must be in PLS format.

LexiconAttributes

lexiconAttributes_languageCode :: Lens' LexiconAttributes (Maybe LanguageCode)

Language code that the lexicon applies to. A lexicon with a language code such as "en" would be applied to all English languages (en-GB, en-US, en-AUS, en-WLS, and so on.

lexiconAttributes_alphabet :: Lens' LexiconAttributes (Maybe Text)

Phonetic alphabet used in the lexicon. Valid values are ipa and x-sampa.

lexiconAttributes_lexiconArn :: Lens' LexiconAttributes (Maybe Text)

Amazon Resource Name (ARN) of the lexicon.

lexiconAttributes_lexemesCount :: Lens' LexiconAttributes (Maybe Int)

Number of lexemes in the lexicon.

lexiconAttributes_lastModified :: Lens' LexiconAttributes (Maybe UTCTime)

Date lexicon was last modified (a timestamp value).

lexiconAttributes_size :: Lens' LexiconAttributes (Maybe Int)

Total size of the lexicon, in characters.

LexiconDescription

lexiconDescription_name :: Lens' LexiconDescription (Maybe Text)

Name of the lexicon.

lexiconDescription_attributes :: Lens' LexiconDescription (Maybe LexiconAttributes)

Provides lexicon metadata.

SynthesisTask

synthesisTask_languageCode :: Lens' SynthesisTask (Maybe LanguageCode)

Optional language code for a synthesis task. This is only necessary if using a bilingual voice, such as Aditi, which can be used for either Indian English (en-IN) or Hindi (hi-IN).

If a bilingual voice is used and no language code is specified, Amazon Polly uses the default language of the bilingual voice. The default language for any voice is the one returned by the DescribeVoices operation for the LanguageCode parameter. For example, if no language code is specified, Aditi will use Indian English rather than Hindi.

synthesisTask_creationTime :: Lens' SynthesisTask (Maybe UTCTime)

Timestamp for the time the synthesis task was started.

synthesisTask_outputUri :: Lens' SynthesisTask (Maybe Text)

Pathway for the output speech file.

synthesisTask_lexiconNames :: Lens' SynthesisTask (Maybe [Text])

List of one or more pronunciation lexicon names you want the service to apply during synthesis. Lexicons are applied only if the language of the lexicon is the same as the language of the voice.

synthesisTask_speechMarkTypes :: Lens' SynthesisTask (Maybe [SpeechMarkType])

The type of speech marks returned for the input text.

synthesisTask_voiceId :: Lens' SynthesisTask (Maybe VoiceId)

Voice ID to use for the synthesis.

synthesisTask_taskId :: Lens' SynthesisTask (Maybe Text)

The Amazon Polly generated identifier for a speech synthesis task.

synthesisTask_textType :: Lens' SynthesisTask (Maybe TextType)

Specifies whether the input text is plain text or SSML. The default value is plain text.

synthesisTask_outputFormat :: Lens' SynthesisTask (Maybe OutputFormat)

The format in which the returned output will be encoded. For audio stream, this will be mp3, ogg_vorbis, or pcm. For speech marks, this will be json.

synthesisTask_sampleRate :: Lens' SynthesisTask (Maybe Text)

The audio frequency specified in Hz.

The valid values for mp3 and ogg_vorbis are "8000", "16000", "22050", and "24000". The default value for standard voices is "22050". The default value for neural voices is "24000".

Valid values for pcm are "8000" and "16000" The default value is "16000".

synthesisTask_taskStatus :: Lens' SynthesisTask (Maybe TaskStatus)

Current status of the individual speech synthesis task.

synthesisTask_engine :: Lens' SynthesisTask (Maybe Engine)

Specifies the engine (standard or neural) for Amazon Polly to use when processing input text for speech synthesis. Using a voice that is not supported for the engine selected will result in an error.

synthesisTask_requestCharacters :: Lens' SynthesisTask (Maybe Int)

Number of billable characters synthesized.

synthesisTask_taskStatusReason :: Lens' SynthesisTask (Maybe Text)

Reason for the current status of a specific speech synthesis task, including errors if the task has failed.

synthesisTask_snsTopicArn :: Lens' SynthesisTask (Maybe Text)

ARN for the SNS topic optionally used for providing status notification for a speech synthesis task.

Voice

voice_languageCode :: Lens' Voice (Maybe LanguageCode)

Language code of the voice.

voice_id :: Lens' Voice (Maybe VoiceId)

Amazon Polly assigned voice ID. This is the ID that you specify when calling the SynthesizeSpeech operation.

voice_name :: Lens' Voice (Maybe Text)

Name of the voice (for example, Salli, Kendra, etc.). This provides a human readable voice name that you might display in your application.

voice_gender :: Lens' Voice (Maybe Gender)

Gender of the voice.

voice_supportedEngines :: Lens' Voice (Maybe [Engine])

Specifies which engines (standard or neural) that are supported by a given voice.

voice_additionalLanguageCodes :: Lens' Voice (Maybe [LanguageCode])

Additional codes for languages available for the specified voice in addition to its default language.

For example, the default language for Aditi is Indian English (en-IN) because it was first used for that language. Since Aditi is bilingual and fluent in both Indian English and Hindi, this parameter would show the code hi-IN.

voice_languageName :: Lens' Voice (Maybe Text)

Human readable name of the language in English.