Network.AWS.Polly.PutLexicon

Description

Stores a pronunciation lexicon in an Amazon Web Services Region. If a lexicon with the same name already exists in the region, it is overwritten by the new lexicon. Lexicon operations have eventual consistency, therefore, it might take some time before the lexicon is available to the SynthesizeSpeech operation.

For more information, see Managing Lexicons.

Creating a Request

data PutLexicon

See: newPutLexicon smart constructor.

Constructors

PutLexicon'

Fields

name :: Text

Name of the lexicon. The name must follow the regular express format [0-9A-Za-z]{1,20}. That is, the name is a case-sensitive alphanumeric string up to 20 characters long.

content :: Sensitive Text

Content of the PLS lexicon as string data.

newPutLexicon

Arguments

:: Text
$sel:name:PutLexicon'

-> Text
$sel:content:PutLexicon'

-> PutLexicon

Create a value of PutLexicon with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:PutLexicon', putLexicon_name - Name of the lexicon. The name must follow the regular express format [0-9A-Za-z]{1,20}. That is, the name is a case-sensitive alphanumeric string up to 20 characters long.

$sel:content:PutLexicon', putLexicon_content - Content of the PLS lexicon as string data.

Request Lenses

putLexicon_name :: Lens' PutLexicon Text

Name of the lexicon. The name must follow the regular express format [0-9A-Za-z]{1,20}. That is, the name is a case-sensitive alphanumeric string up to 20 characters long.

putLexicon_content :: Lens' PutLexicon Text

Content of the PLS lexicon as string data.

Destructuring the Response

data PutLexiconResponse

See: newPutLexiconResponse smart constructor.

Constructors

PutLexiconResponse'

Fields

httpStatus :: Int

The response's http status code.

newPutLexiconResponse

Arguments

:: Int
$sel:httpStatus:PutLexiconResponse'

-> PutLexiconResponse

Create a value of PutLexiconResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:PutLexiconResponse', putLexiconResponse_httpStatus - The response's http status code.

Response Lenses

putLexiconResponse_httpStatus :: Lens' PutLexiconResponse Int

The response's http status code.