Network.AWS.Polly.SynthesizeSpeech

Description

Synthesizes UTF-8 input, plain text or SSML, to a stream of bytes. SSML input must be valid, well-formed SSML. Some alphabets might not be available with all the voices (for example, Cyrillic might not be read at all by English voices) unless phoneme mapping is used. For more information, see How it Works.

Creating a Request

data SynthesizeSpeech

See: newSynthesizeSpeech smart constructor.

Constructors

SynthesizeSpeech'

Fields

languageCode :: Maybe LanguageCode

Optional language code for the Synthesize Speech request. This is only necessary if using a bilingual voice, such as Aditi, which can be used for either Indian English (en-IN) or Hindi (hi-IN).

If a bilingual voice is used and no language code is specified, Amazon Polly uses the default language of the bilingual voice. The default language for any voice is the one returned by the DescribeVoices operation for the LanguageCode parameter. For example, if no language code is specified, Aditi will use Indian English rather than Hindi.

engine :: Maybe Engine

Specifies the engine (standard or neural) for Amazon Polly to use when processing input text for speech synthesis. For information on Amazon Polly voices and which voices are available in standard-only, NTTS-only, and both standard and NTTS formats, see Available Voices.

NTTS-only voices

When using NTTS-only voices such as Kevin (en-US), this parameter is required and must be set to neural. If the engine is not specified, or is set to standard, this will result in an error.

Type: String

Valid Values: standard | neural

Required: Yes

Standard voices

For standard voices, this is not required; the engine parameter defaults to standard. If the engine is not specified, or is set to standard and an NTTS-only voice is selected, this will result in an error.

speechMarkTypes :: Maybe [SpeechMarkType]

The type of speech marks returned for the input text.

sampleRate :: Maybe Text

The audio frequency specified in Hz.

The valid values for mp3 and ogg_vorbis are "8000", "16000", "22050", and "24000". The default value for standard voices is "22050". The default value for neural voices is "24000".

Valid values for pcm are "8000" and "16000" The default value is "16000".

textType :: Maybe TextType

Specifies whether the input text is plain text or SSML. The default value is plain text. For more information, see Using SSML.

lexiconNames :: Maybe [Text]

List of one or more pronunciation lexicon names you want the service to apply during synthesis. Lexicons are applied only if the language of the lexicon is the same as the language of the voice. For information about storing lexicons, see PutLexicon.

outputFormat :: OutputFormat

The format in which the returned output will be encoded. For audio stream, this will be mp3, ogg_vorbis, or pcm. For speech marks, this will be json.

When pcm is used, the content returned is audio/pcm in a signed 16-bit, 1 channel (mono), little-endian format.

text :: Text

Input text to synthesize. If you specify ssml as the TextType, follow the SSML format for the input text.

voiceId :: VoiceId

Voice ID to use for the synthesis. You can get a list of available voice IDs by calling the DescribeVoices operation.

newSynthesizeSpeech

Arguments

:: OutputFormat
$sel:outputFormat:SynthesizeSpeech'

-> Text
$sel:text:SynthesizeSpeech'

-> VoiceId
$sel:voiceId:SynthesizeSpeech'

-> SynthesizeSpeech

Create a value of SynthesizeSpeech with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:SynthesizeSpeech', synthesizeSpeech_languageCode - Optional language code for the Synthesize Speech request. This is only necessary if using a bilingual voice, such as Aditi, which can be used for either Indian English (en-IN) or Hindi (hi-IN).

If a bilingual voice is used and no language code is specified, Amazon Polly uses the default language of the bilingual voice. The default language for any voice is the one returned by the DescribeVoices operation for the LanguageCode parameter. For example, if no language code is specified, Aditi will use Indian English rather than Hindi.

$sel:engine:SynthesizeSpeech', synthesizeSpeech_engine - Specifies the engine (standard or neural) for Amazon Polly to use when processing input text for speech synthesis. For information on Amazon Polly voices and which voices are available in standard-only, NTTS-only, and both standard and NTTS formats, see Available Voices.

NTTS-only voices

When using NTTS-only voices such as Kevin (en-US), this parameter is required and must be set to neural. If the engine is not specified, or is set to standard, this will result in an error.

Type: String

Valid Values: standard | neural

Required: Yes

Standard voices

For standard voices, this is not required; the engine parameter defaults to standard. If the engine is not specified, or is set to standard and an NTTS-only voice is selected, this will result in an error.

$sel:speechMarkTypes:SynthesizeSpeech', synthesizeSpeech_speechMarkTypes - The type of speech marks returned for the input text.

$sel:sampleRate:SynthesizeSpeech', synthesizeSpeech_sampleRate - The audio frequency specified in Hz.

The valid values for mp3 and ogg_vorbis are "8000", "16000", "22050", and "24000". The default value for standard voices is "22050". The default value for neural voices is "24000".

Valid values for pcm are "8000" and "16000" The default value is "16000".

$sel:textType:SynthesizeSpeech', synthesizeSpeech_textType - Specifies whether the input text is plain text or SSML. The default value is plain text. For more information, see Using SSML.

$sel:lexiconNames:SynthesizeSpeech', synthesizeSpeech_lexiconNames - List of one or more pronunciation lexicon names you want the service to apply during synthesis. Lexicons are applied only if the language of the lexicon is the same as the language of the voice. For information about storing lexicons, see PutLexicon.

$sel:outputFormat:SynthesizeSpeech', synthesizeSpeech_outputFormat - The format in which the returned output will be encoded. For audio stream, this will be mp3, ogg_vorbis, or pcm. For speech marks, this will be json.

When pcm is used, the content returned is audio/pcm in a signed 16-bit, 1 channel (mono), little-endian format.

$sel:text:SynthesizeSpeech', synthesizeSpeech_text - Input text to synthesize. If you specify ssml as the TextType, follow the SSML format for the input text.

$sel:voiceId:SynthesizeSpeech', synthesizeSpeech_voiceId - Voice ID to use for the synthesis. You can get a list of available voice IDs by calling the DescribeVoices operation.

Request Lenses

synthesizeSpeech_languageCode :: Lens' SynthesizeSpeech (Maybe LanguageCode)

Optional language code for the Synthesize Speech request. This is only necessary if using a bilingual voice, such as Aditi, which can be used for either Indian English (en-IN) or Hindi (hi-IN).

If a bilingual voice is used and no language code is specified, Amazon Polly uses the default language of the bilingual voice. The default language for any voice is the one returned by the DescribeVoices operation for the LanguageCode parameter. For example, if no language code is specified, Aditi will use Indian English rather than Hindi.

synthesizeSpeech_engine :: Lens' SynthesizeSpeech (Maybe Engine)

Specifies the engine (standard or neural) for Amazon Polly to use when processing input text for speech synthesis. For information on Amazon Polly voices and which voices are available in standard-only, NTTS-only, and both standard and NTTS formats, see Available Voices.

NTTS-only voices

When using NTTS-only voices such as Kevin (en-US), this parameter is required and must be set to neural. If the engine is not specified, or is set to standard, this will result in an error.

Type: String

Valid Values: standard | neural

Required: Yes

Standard voices

For standard voices, this is not required; the engine parameter defaults to standard. If the engine is not specified, or is set to standard and an NTTS-only voice is selected, this will result in an error.

synthesizeSpeech_speechMarkTypes :: Lens' SynthesizeSpeech (Maybe [SpeechMarkType])

The type of speech marks returned for the input text.

synthesizeSpeech_sampleRate :: Lens' SynthesizeSpeech (Maybe Text)

The audio frequency specified in Hz.

The valid values for mp3 and ogg_vorbis are "8000", "16000", "22050", and "24000". The default value for standard voices is "22050". The default value for neural voices is "24000".

Valid values for pcm are "8000" and "16000" The default value is "16000".

synthesizeSpeech_textType :: Lens' SynthesizeSpeech (Maybe TextType)

Specifies whether the input text is plain text or SSML. The default value is plain text. For more information, see Using SSML.

synthesizeSpeech_lexiconNames :: Lens' SynthesizeSpeech (Maybe [Text])

List of one or more pronunciation lexicon names you want the service to apply during synthesis. Lexicons are applied only if the language of the lexicon is the same as the language of the voice. For information about storing lexicons, see PutLexicon.

synthesizeSpeech_outputFormat :: Lens' SynthesizeSpeech OutputFormat

The format in which the returned output will be encoded. For audio stream, this will be mp3, ogg_vorbis, or pcm. For speech marks, this will be json.

When pcm is used, the content returned is audio/pcm in a signed 16-bit, 1 channel (mono), little-endian format.

synthesizeSpeech_text :: Lens' SynthesizeSpeech Text

Input text to synthesize. If you specify ssml as the TextType, follow the SSML format for the input text.

synthesizeSpeech_voiceId :: Lens' SynthesizeSpeech VoiceId

Voice ID to use for the synthesis. You can get a list of available voice IDs by calling the DescribeVoices operation.

Destructuring the Response

data SynthesizeSpeechResponse

See: newSynthesizeSpeechResponse smart constructor.

Constructors

SynthesizeSpeechResponse'

Fields

requestCharacters :: Maybe Int

Number of characters synthesized.

contentType :: Maybe Text

Specifies the type audio stream. This should reflect the OutputFormat parameter in your request.

If you request mp3 as the OutputFormat, the ContentType returned is audio/mpeg.

If you request ogg_vorbis as the OutputFormat, the ContentType returned is audio/ogg.

If you request pcm as the OutputFormat, the ContentType returned is audio/pcm in a signed 16-bit, 1 channel (mono), little-endian format.

If you request json as the OutputFormat, the ContentType returned is audio/json.

httpStatus :: Int

The response's http status code.

audioStream :: ResponseBody

Stream containing the synthesized speech.

newSynthesizeSpeechResponse

Arguments

:: Int
$sel:httpStatus:SynthesizeSpeechResponse'

-> ResponseBody
$sel:audioStream:SynthesizeSpeechResponse'

-> SynthesizeSpeechResponse

Create a value of SynthesizeSpeechResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestCharacters:SynthesizeSpeechResponse', synthesizeSpeechResponse_requestCharacters - Number of characters synthesized.

$sel:contentType:SynthesizeSpeechResponse', synthesizeSpeechResponse_contentType - Specifies the type audio stream. This should reflect the OutputFormat parameter in your request.

If you request mp3 as the OutputFormat, the ContentType returned is audio/mpeg.

If you request ogg_vorbis as the OutputFormat, the ContentType returned is audio/ogg.

If you request pcm as the OutputFormat, the ContentType returned is audio/pcm in a signed 16-bit, 1 channel (mono), little-endian format.

If you request json as the OutputFormat, the ContentType returned is audio/json.

$sel:httpStatus:SynthesizeSpeechResponse', synthesizeSpeechResponse_httpStatus - The response's http status code.

$sel:audioStream:SynthesizeSpeechResponse', synthesizeSpeechResponse_audioStream - Stream containing the synthesized speech.

Response Lenses

synthesizeSpeechResponse_requestCharacters :: Lens' SynthesizeSpeechResponse (Maybe Int)

Number of characters synthesized.

synthesizeSpeechResponse_contentType :: Lens' SynthesizeSpeechResponse (Maybe Text)

Specifies the type audio stream. This should reflect the OutputFormat parameter in your request.

If you request mp3 as the OutputFormat, the ContentType returned is audio/mpeg.

If you request ogg_vorbis as the OutputFormat, the ContentType returned is audio/ogg.

If you request pcm as the OutputFormat, the ContentType returned is audio/pcm in a signed 16-bit, 1 channel (mono), little-endian format.

If you request json as the OutputFormat, the ContentType returned is audio/json.

synthesizeSpeechResponse_httpStatus :: Lens' SynthesizeSpeechResponse Int

The response's http status code.

synthesizeSpeechResponse_audioStream :: Lens' SynthesizeSpeechResponse ResponseBody

Stream containing the synthesized speech.