Network.AWS.Polly.Types

Description

Service Configuration

defaultService :: Service

API version 2016-06-10 of the Amazon Polly SDK configuration.

Errors

_InvalidSsmlException :: AsError a => Getting (First ServiceError) a ServiceError

The SSML you provided is invalid. Verify the SSML syntax, spelling of tags and values, and then try again.

_InvalidLexiconException :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Polly can't find the specified lexicon. Verify that the lexicon's name is spelled correctly, and then try again.

_InvalidS3KeyException :: AsError a => Getting (First ServiceError) a ServiceError

The provided Amazon S3 key prefix is invalid. Please provide a valid S3 object key name.

_LexiconNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Polly can't find the specified lexicon. This could be caused by a lexicon that is missing, its name is misspelled or specifying a lexicon that is in a different region.

Verify that the lexicon exists, is in the region (see ListLexicons) and that you spelled its name is spelled correctly. Then try again.

_MaxLexemeLengthExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The maximum size of the lexeme would be exceeded by this operation.

_TextLengthExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The value of the "Text" parameter is longer than the accepted limits. For the SynthesizeSpeech API, the limit for input text is a maximum of 6000 characters total, of which no more than 3000 can be billed characters. For the StartSpeechSynthesisTask API, the maximum is 200,000 characters, of which no more than 100,000 can be billed characters. SSML tags are not counted as billed characters.

_SynthesisTaskNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The Speech Synthesis task with requested Task ID cannot be found.

_InvalidSampleRateException :: AsError a => Getting (First ServiceError) a ServiceError

The specified sample rate is not valid.

_ServiceFailureException :: AsError a => Getting (First ServiceError) a ServiceError

An unknown condition has caused a service failure.

_MarksNotSupportedForFormatException :: AsError a => Getting (First ServiceError) a ServiceError

Speech marks are not supported for the OutputFormat selected. Speech marks are only available for content in json format.

_MaxLexiconsNumberExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The maximum number of lexicons would be exceeded by this operation.

_UnsupportedPlsAlphabetException :: AsError a => Getting (First ServiceError) a ServiceError

The alphabet specified by the lexicon is not a supported alphabet. Valid values are x-sampa and ipa.

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The NextToken is invalid. Verify that it's spelled correctly, and then try again.

_UnsupportedPlsLanguageException :: AsError a => Getting (First ServiceError) a ServiceError

The language specified in the lexicon is unsupported. For a list of supported languages, see Lexicon Attributes.

_InvalidSnsTopicArnException :: AsError a => Getting (First ServiceError) a ServiceError

The provided SNS topic ARN is invalid. Please provide a valid SNS topic ARN and try again.

_InvalidTaskIdException :: AsError a => Getting (First ServiceError) a ServiceError

The provided Task ID is not valid. Please provide a valid Task ID and try again.

_LanguageNotSupportedException :: AsError a => Getting (First ServiceError) a ServiceError

The language specified is not currently supported by Amazon Polly in this capacity.

_LexiconSizeExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The maximum size of the specified lexicon would be exceeded by this operation.

_InvalidS3BucketException :: AsError a => Getting (First ServiceError) a ServiceError

The provided Amazon S3 bucket name is invalid. Please check your input with S3 bucket naming requirements and try again.

_SsmlMarksNotSupportedForTextTypeException :: AsError a => Getting (First ServiceError) a ServiceError

SSML speech marks are not supported for plain text-type input.

_EngineNotSupportedException :: AsError a => Getting (First ServiceError) a ServiceError

This engine is not compatible with the voice that you have designated. Choose a new voice that is compatible with the engine or change the engine and restart the operation.

Engine

newtype Engine

Constructors

Engine'

Fields

fromEngine :: Text

Bundled Patterns

pattern Engine_Neural :: Engine

pattern Engine_Standard :: Engine

Gender

newtype Gender

Constructors

Gender'

Fields

fromGender :: Text

Bundled Patterns

pattern Gender_Female :: Gender

pattern Gender_Male :: Gender

LanguageCode

newtype LanguageCode

Constructors

LanguageCode'

Fields

fromLanguageCode :: Text

Bundled Patterns

pattern LanguageCode_Arb :: LanguageCode

pattern LanguageCode_Cmn_CN :: LanguageCode

pattern LanguageCode_Cy_GB :: LanguageCode

pattern LanguageCode_Da_DK :: LanguageCode

pattern LanguageCode_De_DE :: LanguageCode

pattern LanguageCode_En_AU :: LanguageCode

pattern LanguageCode_En_GB :: LanguageCode

pattern LanguageCode_En_GB_WLS :: LanguageCode

pattern LanguageCode_En_IN :: LanguageCode

pattern LanguageCode_En_NZ :: LanguageCode

pattern LanguageCode_En_US :: LanguageCode

pattern LanguageCode_En_ZA :: LanguageCode

pattern LanguageCode_Es_ES :: LanguageCode

pattern LanguageCode_Es_MX :: LanguageCode

pattern LanguageCode_Es_US :: LanguageCode

pattern LanguageCode_Fr_CA :: LanguageCode

pattern LanguageCode_Fr_FR :: LanguageCode

pattern LanguageCode_Hi_IN :: LanguageCode

pattern LanguageCode_Is_IS :: LanguageCode

pattern LanguageCode_It_IT :: LanguageCode

pattern LanguageCode_Ja_JP :: LanguageCode

pattern LanguageCode_Ko_KR :: LanguageCode

pattern LanguageCode_Nb_NO :: LanguageCode

pattern LanguageCode_Nl_NL :: LanguageCode

pattern LanguageCode_Pl_PL :: LanguageCode

pattern LanguageCode_Pt_BR :: LanguageCode

pattern LanguageCode_Pt_PT :: LanguageCode

pattern LanguageCode_Ro_RO :: LanguageCode

pattern LanguageCode_Ru_RU :: LanguageCode

pattern LanguageCode_Sv_SE :: LanguageCode

pattern LanguageCode_Tr_TR :: LanguageCode

OutputFormat

newtype OutputFormat

Constructors

OutputFormat'

Fields

fromOutputFormat :: Text

Bundled Patterns

pattern OutputFormat_Json :: OutputFormat

pattern OutputFormat_Mp3 :: OutputFormat

pattern OutputFormat_Ogg_vorbis :: OutputFormat

pattern OutputFormat_Pcm :: OutputFormat

SpeechMarkType

newtype SpeechMarkType

Constructors

SpeechMarkType'

Fields

fromSpeechMarkType :: Text

Bundled Patterns

pattern SpeechMarkType_Sentence :: SpeechMarkType

pattern SpeechMarkType_Ssml :: SpeechMarkType

pattern SpeechMarkType_Viseme :: SpeechMarkType

pattern SpeechMarkType_Word :: SpeechMarkType

TaskStatus

newtype TaskStatus

Constructors

TaskStatus'

Fields

fromTaskStatus :: Text

Bundled Patterns

pattern TaskStatus_Completed :: TaskStatus

pattern TaskStatus_Failed :: TaskStatus

pattern TaskStatus_InProgress :: TaskStatus

pattern TaskStatus_Scheduled :: TaskStatus

TextType

newtype TextType

Constructors

TextType'

Fields

fromTextType :: Text

Bundled Patterns

pattern TextType_Ssml :: TextType

pattern TextType_Text :: TextType

VoiceId

newtype VoiceId

Constructors

VoiceId'

Fields

fromVoiceId :: Text

Bundled Patterns

pattern VoiceId_Aditi :: VoiceId

pattern VoiceId_Amy :: VoiceId

pattern VoiceId_Aria :: VoiceId

pattern VoiceId_Astrid :: VoiceId

pattern VoiceId_Ayanda :: VoiceId

pattern VoiceId_Bianca :: VoiceId

pattern VoiceId_Brian :: VoiceId

pattern VoiceId_Camila :: VoiceId

pattern VoiceId_Carla :: VoiceId

pattern VoiceId_Carmen :: VoiceId

pattern VoiceId_Celine :: VoiceId

pattern VoiceId_Chantal :: VoiceId

pattern VoiceId_Conchita :: VoiceId

pattern VoiceId_Cristiano :: VoiceId

pattern VoiceId_Dora :: VoiceId

pattern VoiceId_Emma :: VoiceId

pattern VoiceId_Enrique :: VoiceId

pattern VoiceId_Ewa :: VoiceId

pattern VoiceId_Filiz :: VoiceId

pattern VoiceId_Gabrielle :: VoiceId

pattern VoiceId_Geraint :: VoiceId

pattern VoiceId_Giorgio :: VoiceId

pattern VoiceId_Gwyneth :: VoiceId

pattern VoiceId_Hans :: VoiceId

pattern VoiceId_Ines :: VoiceId

pattern VoiceId_Ivy :: VoiceId

pattern VoiceId_Jacek :: VoiceId

pattern VoiceId_Jan :: VoiceId

pattern VoiceId_Joanna :: VoiceId

pattern VoiceId_Joey :: VoiceId

pattern VoiceId_Justin :: VoiceId

pattern VoiceId_Karl :: VoiceId

pattern VoiceId_Kendra :: VoiceId

pattern VoiceId_Kevin :: VoiceId

pattern VoiceId_Kimberly :: VoiceId

pattern VoiceId_Lea :: VoiceId

pattern VoiceId_Liv :: VoiceId

pattern VoiceId_Lotte :: VoiceId

pattern VoiceId_Lucia :: VoiceId

pattern VoiceId_Lupe :: VoiceId

pattern VoiceId_Mads :: VoiceId

pattern VoiceId_Maja :: VoiceId

pattern VoiceId_Marlene :: VoiceId

pattern VoiceId_Mathieu :: VoiceId

pattern VoiceId_Matthew :: VoiceId

pattern VoiceId_Maxim :: VoiceId

pattern VoiceId_Mia :: VoiceId

pattern VoiceId_Miguel :: VoiceId

pattern VoiceId_Mizuki :: VoiceId

pattern VoiceId_Naja :: VoiceId

pattern VoiceId_Nicole :: VoiceId

pattern VoiceId_Olivia :: VoiceId

pattern VoiceId_Penelope :: VoiceId

pattern VoiceId_Raveena :: VoiceId

pattern VoiceId_Ricardo :: VoiceId

pattern VoiceId_Ruben :: VoiceId

pattern VoiceId_Russell :: VoiceId

pattern VoiceId_Salli :: VoiceId

pattern VoiceId_Seoyeon :: VoiceId

pattern VoiceId_Takumi :: VoiceId

pattern VoiceId_Tatyana :: VoiceId

pattern VoiceId_Vicki :: VoiceId

pattern VoiceId_Vitoria :: VoiceId

pattern VoiceId_Zeina :: VoiceId

pattern VoiceId_Zhiyu :: VoiceId

Lexicon

data Lexicon

Provides lexicon name and lexicon content in string format. For more information, see Pronunciation Lexicon Specification (PLS) Version 1.0.

See: newLexicon smart constructor.

Constructors

Lexicon'

Fields

name :: Maybe Text

Name of the lexicon.

content :: Maybe (Sensitive Text)

Lexicon content in string format. The content of a lexicon must be in PLS format.

newLexicon :: Lexicon

Create a value of Lexicon with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:Lexicon', lexicon_name - Name of the lexicon.

$sel:content:Lexicon', lexicon_content - Lexicon content in string format. The content of a lexicon must be in PLS format.

lexicon_name :: Lens' Lexicon (Maybe Text)

Name of the lexicon.

lexicon_content :: Lens' Lexicon (Maybe Text)

Lexicon content in string format. The content of a lexicon must be in PLS format.

LexiconAttributes

data LexiconAttributes

Contains metadata describing the lexicon such as the number of lexemes, language code, and so on. For more information, see Managing Lexicons.

See: newLexiconAttributes smart constructor.

Constructors

LexiconAttributes'

Fields

languageCode :: Maybe LanguageCode

Language code that the lexicon applies to. A lexicon with a language code such as "en" would be applied to all English languages (en-GB, en-US, en-AUS, en-WLS, and so on.

alphabet :: Maybe Text

Phonetic alphabet used in the lexicon. Valid values are ipa and x-sampa.

lexiconArn :: Maybe Text

Amazon Resource Name (ARN) of the lexicon.

lexemesCount :: Maybe Int

Number of lexemes in the lexicon.

lastModified :: Maybe POSIX

Date lexicon was last modified (a timestamp value).

size :: Maybe Int

Total size of the lexicon, in characters.

newLexiconAttributes :: LexiconAttributes

Create a value of LexiconAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:LexiconAttributes', lexiconAttributes_languageCode - Language code that the lexicon applies to. A lexicon with a language code such as "en" would be applied to all English languages (en-GB, en-US, en-AUS, en-WLS, and so on.

$sel:alphabet:LexiconAttributes', lexiconAttributes_alphabet - Phonetic alphabet used in the lexicon. Valid values are ipa and x-sampa.

$sel:lexiconArn:LexiconAttributes', lexiconAttributes_lexiconArn - Amazon Resource Name (ARN) of the lexicon.

$sel:lexemesCount:LexiconAttributes', lexiconAttributes_lexemesCount - Number of lexemes in the lexicon.

$sel:lastModified:LexiconAttributes', lexiconAttributes_lastModified - Date lexicon was last modified (a timestamp value).

$sel:size:LexiconAttributes', lexiconAttributes_size - Total size of the lexicon, in characters.

lexiconAttributes_languageCode :: Lens' LexiconAttributes (Maybe LanguageCode)

Language code that the lexicon applies to. A lexicon with a language code such as "en" would be applied to all English languages (en-GB, en-US, en-AUS, en-WLS, and so on.

lexiconAttributes_alphabet :: Lens' LexiconAttributes (Maybe Text)

Phonetic alphabet used in the lexicon. Valid values are ipa and x-sampa.

lexiconAttributes_lexiconArn :: Lens' LexiconAttributes (Maybe Text)

Amazon Resource Name (ARN) of the lexicon.

lexiconAttributes_lexemesCount :: Lens' LexiconAttributes (Maybe Int)

Number of lexemes in the lexicon.

lexiconAttributes_lastModified :: Lens' LexiconAttributes (Maybe UTCTime)

Date lexicon was last modified (a timestamp value).

lexiconAttributes_size :: Lens' LexiconAttributes (Maybe Int)

Total size of the lexicon, in characters.

LexiconDescription

data LexiconDescription

Describes the content of the lexicon.

See: newLexiconDescription smart constructor.

Constructors

LexiconDescription'

Fields

name :: Maybe Text

Name of the lexicon.

attributes :: Maybe LexiconAttributes

Provides lexicon metadata.

newLexiconDescription :: LexiconDescription

Create a value of LexiconDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:LexiconDescription', lexiconDescription_name - Name of the lexicon.

$sel:attributes:LexiconDescription', lexiconDescription_attributes - Provides lexicon metadata.

lexiconDescription_name :: Lens' LexiconDescription (Maybe Text)

Name of the lexicon.

lexiconDescription_attributes :: Lens' LexiconDescription (Maybe LexiconAttributes)

Provides lexicon metadata.

SynthesisTask

data SynthesisTask

SynthesisTask object that provides information about a speech synthesis task.

See: newSynthesisTask smart constructor.

Constructors

SynthesisTask'

Fields

languageCode :: Maybe LanguageCode

Optional language code for a synthesis task. This is only necessary if using a bilingual voice, such as Aditi, which can be used for either Indian English (en-IN) or Hindi (hi-IN).

If a bilingual voice is used and no language code is specified, Amazon Polly uses the default language of the bilingual voice. The default language for any voice is the one returned by the DescribeVoices operation for the LanguageCode parameter. For example, if no language code is specified, Aditi will use Indian English rather than Hindi.

creationTime :: Maybe POSIX

Timestamp for the time the synthesis task was started.

outputUri :: Maybe Text

Pathway for the output speech file.

lexiconNames :: Maybe [Text]

List of one or more pronunciation lexicon names you want the service to apply during synthesis. Lexicons are applied only if the language of the lexicon is the same as the language of the voice.

speechMarkTypes :: Maybe [SpeechMarkType]

The type of speech marks returned for the input text.

voiceId :: Maybe VoiceId

Voice ID to use for the synthesis.

taskId :: Maybe Text

The Amazon Polly generated identifier for a speech synthesis task.

textType :: Maybe TextType

Specifies whether the input text is plain text or SSML. The default value is plain text.

outputFormat :: Maybe OutputFormat

The format in which the returned output will be encoded. For audio stream, this will be mp3, ogg_vorbis, or pcm. For speech marks, this will be json.

sampleRate :: Maybe Text

The audio frequency specified in Hz.

The valid values for mp3 and ogg_vorbis are "8000", "16000", "22050", and "24000". The default value for standard voices is "22050". The default value for neural voices is "24000".

Valid values for pcm are "8000" and "16000" The default value is "16000".

taskStatus :: Maybe TaskStatus

Current status of the individual speech synthesis task.

engine :: Maybe Engine

Specifies the engine (standard or neural) for Amazon Polly to use when processing input text for speech synthesis. Using a voice that is not supported for the engine selected will result in an error.

requestCharacters :: Maybe Int

Number of billable characters synthesized.

taskStatusReason :: Maybe Text

Reason for the current status of a specific speech synthesis task, including errors if the task has failed.

snsTopicArn :: Maybe Text

ARN for the SNS topic optionally used for providing status notification for a speech synthesis task.

newSynthesisTask :: SynthesisTask

Create a value of SynthesisTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:SynthesisTask', synthesisTask_languageCode - Optional language code for a synthesis task. This is only necessary if using a bilingual voice, such as Aditi, which can be used for either Indian English (en-IN) or Hindi (hi-IN).

If a bilingual voice is used and no language code is specified, Amazon Polly uses the default language of the bilingual voice. The default language for any voice is the one returned by the DescribeVoices operation for the LanguageCode parameter. For example, if no language code is specified, Aditi will use Indian English rather than Hindi.

$sel:creationTime:SynthesisTask', synthesisTask_creationTime - Timestamp for the time the synthesis task was started.

$sel:outputUri:SynthesisTask', synthesisTask_outputUri - Pathway for the output speech file.

$sel:lexiconNames:SynthesisTask', synthesisTask_lexiconNames - List of one or more pronunciation lexicon names you want the service to apply during synthesis. Lexicons are applied only if the language of the lexicon is the same as the language of the voice.

$sel:speechMarkTypes:SynthesisTask', synthesisTask_speechMarkTypes - The type of speech marks returned for the input text.

$sel:voiceId:SynthesisTask', synthesisTask_voiceId - Voice ID to use for the synthesis.

$sel:taskId:SynthesisTask', synthesisTask_taskId - The Amazon Polly generated identifier for a speech synthesis task.

$sel:textType:SynthesisTask', synthesisTask_textType - Specifies whether the input text is plain text or SSML. The default value is plain text.

$sel:outputFormat:SynthesisTask', synthesisTask_outputFormat - The format in which the returned output will be encoded. For audio stream, this will be mp3, ogg_vorbis, or pcm. For speech marks, this will be json.

$sel:sampleRate:SynthesisTask', synthesisTask_sampleRate - The audio frequency specified in Hz.

The valid values for mp3 and ogg_vorbis are "8000", "16000", "22050", and "24000". The default value for standard voices is "22050". The default value for neural voices is "24000".

Valid values for pcm are "8000" and "16000" The default value is "16000".

$sel:taskStatus:SynthesisTask', synthesisTask_taskStatus - Current status of the individual speech synthesis task.

$sel:engine:SynthesisTask', synthesisTask_engine - Specifies the engine (standard or neural) for Amazon Polly to use when processing input text for speech synthesis. Using a voice that is not supported for the engine selected will result in an error.

$sel:requestCharacters:SynthesisTask', synthesisTask_requestCharacters - Number of billable characters synthesized.

$sel:taskStatusReason:SynthesisTask', synthesisTask_taskStatusReason - Reason for the current status of a specific speech synthesis task, including errors if the task has failed.

$sel:snsTopicArn:SynthesisTask', synthesisTask_snsTopicArn - ARN for the SNS topic optionally used for providing status notification for a speech synthesis task.

synthesisTask_languageCode :: Lens' SynthesisTask (Maybe LanguageCode)

Optional language code for a synthesis task. This is only necessary if using a bilingual voice, such as Aditi, which can be used for either Indian English (en-IN) or Hindi (hi-IN).

If a bilingual voice is used and no language code is specified, Amazon Polly uses the default language of the bilingual voice. The default language for any voice is the one returned by the DescribeVoices operation for the LanguageCode parameter. For example, if no language code is specified, Aditi will use Indian English rather than Hindi.

synthesisTask_creationTime :: Lens' SynthesisTask (Maybe UTCTime)

Timestamp for the time the synthesis task was started.

synthesisTask_outputUri :: Lens' SynthesisTask (Maybe Text)

Pathway for the output speech file.

synthesisTask_lexiconNames :: Lens' SynthesisTask (Maybe [Text])

List of one or more pronunciation lexicon names you want the service to apply during synthesis. Lexicons are applied only if the language of the lexicon is the same as the language of the voice.

synthesisTask_speechMarkTypes :: Lens' SynthesisTask (Maybe [SpeechMarkType])

The type of speech marks returned for the input text.

synthesisTask_voiceId :: Lens' SynthesisTask (Maybe VoiceId)

Voice ID to use for the synthesis.

synthesisTask_taskId :: Lens' SynthesisTask (Maybe Text)

The Amazon Polly generated identifier for a speech synthesis task.

synthesisTask_textType :: Lens' SynthesisTask (Maybe TextType)

Specifies whether the input text is plain text or SSML. The default value is plain text.

synthesisTask_outputFormat :: Lens' SynthesisTask (Maybe OutputFormat)

The format in which the returned output will be encoded. For audio stream, this will be mp3, ogg_vorbis, or pcm. For speech marks, this will be json.

synthesisTask_sampleRate :: Lens' SynthesisTask (Maybe Text)

The audio frequency specified in Hz.

The valid values for mp3 and ogg_vorbis are "8000", "16000", "22050", and "24000". The default value for standard voices is "22050". The default value for neural voices is "24000".

Valid values for pcm are "8000" and "16000" The default value is "16000".

synthesisTask_taskStatus :: Lens' SynthesisTask (Maybe TaskStatus)

Current status of the individual speech synthesis task.

synthesisTask_engine :: Lens' SynthesisTask (Maybe Engine)

Specifies the engine (standard or neural) for Amazon Polly to use when processing input text for speech synthesis. Using a voice that is not supported for the engine selected will result in an error.

synthesisTask_requestCharacters :: Lens' SynthesisTask (Maybe Int)

Number of billable characters synthesized.

synthesisTask_taskStatusReason :: Lens' SynthesisTask (Maybe Text)

Reason for the current status of a specific speech synthesis task, including errors if the task has failed.

synthesisTask_snsTopicArn :: Lens' SynthesisTask (Maybe Text)

ARN for the SNS topic optionally used for providing status notification for a speech synthesis task.

Voice

data Voice

Description of the voice.

See: newVoice smart constructor.

Constructors

Voice'

Fields

languageCode :: Maybe LanguageCode

Language code of the voice.

id :: Maybe VoiceId

Amazon Polly assigned voice ID. This is the ID that you specify when calling the SynthesizeSpeech operation.

name :: Maybe Text

Name of the voice (for example, Salli, Kendra, etc.). This provides a human readable voice name that you might display in your application.

gender :: Maybe Gender

Gender of the voice.

supportedEngines :: Maybe [Engine]

Specifies which engines (standard or neural) that are supported by a given voice.

additionalLanguageCodes :: Maybe [LanguageCode]

Additional codes for languages available for the specified voice in addition to its default language.

For example, the default language for Aditi is Indian English (en-IN) because it was first used for that language. Since Aditi is bilingual and fluent in both Indian English and Hindi, this parameter would show the code hi-IN.

languageName :: Maybe Text

Human readable name of the language in English.

newVoice :: Voice

Create a value of Voice with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:languageCode:Voice', voice_languageCode - Language code of the voice.

$sel:id:Voice', voice_id - Amazon Polly assigned voice ID. This is the ID that you specify when calling the SynthesizeSpeech operation.

$sel:name:Voice', voice_name - Name of the voice (for example, Salli, Kendra, etc.). This provides a human readable voice name that you might display in your application.

$sel:gender:Voice', voice_gender - Gender of the voice.

$sel:supportedEngines:Voice', voice_supportedEngines - Specifies which engines (standard or neural) that are supported by a given voice.

$sel:additionalLanguageCodes:Voice', voice_additionalLanguageCodes - Additional codes for languages available for the specified voice in addition to its default language.

For example, the default language for Aditi is Indian English (en-IN) because it was first used for that language. Since Aditi is bilingual and fluent in both Indian English and Hindi, this parameter would show the code hi-IN.

$sel:languageName:Voice', voice_languageName - Human readable name of the language in English.

voice_languageCode :: Lens' Voice (Maybe LanguageCode)

Language code of the voice.

voice_id :: Lens' Voice (Maybe VoiceId)

Amazon Polly assigned voice ID. This is the ID that you specify when calling the SynthesizeSpeech operation.

voice_name :: Lens' Voice (Maybe Text)

Name of the voice (for example, Salli, Kendra, etc.). This provides a human readable voice name that you might display in your application.

voice_gender :: Lens' Voice (Maybe Gender)

Gender of the voice.

voice_supportedEngines :: Lens' Voice (Maybe [Engine])

Specifies which engines (standard or neural) that are supported by a given voice.

voice_additionalLanguageCodes :: Lens' Voice (Maybe [LanguageCode])

Additional codes for languages available for the specified voice in addition to its default language.

For example, the default language for Aditi is Indian English (en-IN) because it was first used for that language. Since Aditi is bilingual and fluent in both Indian English and Hindi, this parameter would show the code hi-IN.

voice_languageName :: Lens' Voice (Maybe Text)

Human readable name of the language in English.