Network.AWS.Prelude

Re-exported

Deriving

class Generic a

Representable types of kind *. This class is derivable in GHC with the DeriveGeneric flag on.

Minimal complete definition

from, to

class Typeable a

The class Typeable allows a concrete representation of a type to be calculated.

Minimal complete definition

typeRep#

Primitives

data Base64

data ByteString :: *

A space-efficient representation of a Word8 vector, supporting many efficient operations.

A ByteString contains 8-bit bytes, or by using the operations from Data.ByteString.Char8 it can be interpreted as containing 8-bit characters.

class (Typeable * e, Show e) => Exception e

Any type that you wish to throw or catch as an exception must be an instance of the Exception class. The simplest case is a new exception type directly below the root:

data MyException = ThisException | ThatException
    deriving (Show, Typeable)

instance Exception MyException

The default method definitions in the Exception class do what we need in this case. You can now throw and catch ThisException and ThatException as exceptions:

*Main> throw ThisException `catch` \e -> putStrLn ("Caught " ++ show (e :: MyException))
Caught ThisException

In more complicated examples, you may wish to define a whole hierarchy of exceptions:

---------------------------------------------------------------------
-- Make the root exception type for all the exceptions in a compiler

data SomeCompilerException = forall e . Exception e => SomeCompilerException e
    deriving Typeable

instance Show SomeCompilerException where
    show (SomeCompilerException e) = show e

instance Exception SomeCompilerException

compilerExceptionToException :: Exception e => e -> SomeException
compilerExceptionToException = toException . SomeCompilerException

compilerExceptionFromException :: Exception e => SomeException -> Maybe e
compilerExceptionFromException x = do
    SomeCompilerException a <- fromException x
    cast a

---------------------------------------------------------------------
-- Make a subhierarchy for exceptions in the frontend of the compiler

data SomeFrontendException = forall e . Exception e => SomeFrontendException e
    deriving Typeable

instance Show SomeFrontendException where
    show (SomeFrontendException e) = show e

instance Exception SomeFrontendException where
    toException = compilerExceptionToException
    fromException = compilerExceptionFromException

frontendExceptionToException :: Exception e => e -> SomeException
frontendExceptionToException = toException . SomeFrontendException

frontendExceptionFromException :: Exception e => SomeException -> Maybe e
frontendExceptionFromException x = do
    SomeFrontendException a <- fromException x
    cast a

---------------------------------------------------------------------
-- Make an exception type for a particular frontend compiler exception

data MismatchedParentheses = MismatchedParentheses
    deriving (Typeable, Show)

instance Exception MismatchedParentheses where
    toException   = frontendExceptionToException
    fromException = frontendExceptionFromException

We can now catch a MismatchedParentheses exception as MismatchedParentheses, SomeFrontendException or SomeCompilerException, but not other types, e.g. IOException:

*Main> throw MismatchedParentheses catch e -> putStrLn ("Caught " ++ show (e :: MismatchedParentheses))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses catch e -> putStrLn ("Caught " ++ show (e :: SomeFrontendException))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses catch e -> putStrLn ("Caught " ++ show (e :: SomeCompilerException))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses catch e -> putStrLn ("Caught " ++ show (e :: IOException))
*** Exception: MismatchedParentheses

data HashMap k v :: * -> * -> *

data HttpException :: *

data RequestBody :: *

data Response body :: * -> *

data Text :: *

Classes

class FromJSON a where

Minimal complete definition

Nothing

Methods

parseJSON :: Value -> Parser a

class ToJSON a where

Minimal complete definition

Nothing

Methods

toJSON :: a -> Value

class AWSError a where

Methods

awsError :: a -> Error

class AWSError a => AWSServiceError a where

Methods

serviceError :: String -> a

clientError :: HttpException -> a

serializerError :: String -> a

class (AWSSigner (Sg a), AWSServiceError (Er a)) => AWSService a where

Associated Types

type Sg a :: *

data Er a :: *

Methods

service :: Service' a

class (AWSService (Sv a), AWSSigner (Sg (Sv a))) => AWSRequest a where

Associated Types

type Sv a :: *

type Rs a :: *

Methods

request :: a -> Request a

response :: Monad m => a -> Either HttpException (ClientResponse m) -> m (Either (Er (Sv a)) (Rs a))

class AWSRequest a => AWSPager a where

Methods

next :: a -> Rs a -> Maybe a

Shared

data Service' a

Constructors

Service'

Fields

_svcEndpoint :: !Endpoint'

_svcPrefix :: ByteString

_svcVersion :: ByteString

_svcTarget :: Maybe ByteString

newtype Action

Constructors

Action Text

newtype BucketName

Constructors

BucketName Text

newtype ObjectKey

Constructors

ObjectKey Text

newtype ObjectVersionId

Constructors

ObjectVersionId Text

newtype ETag

Constructors

ETag Text

data Switch a

Constructors

Enabled

Disabled

data RecordType

Constructors

A

AAAA

CNAME

MX

NS

PTR

SOA

SPF

SRV

TXT

data Endpoint'

Constructors

Global

Regional

Custom ByteString