Network.AWS.Pricing.DescribeServices

Description

Returns the metadata for one service or a list of the metadata for all services. Use this without a service code to get the service codes for all services. Use it with a service code, such as AmazonEC2, to get information specific to that service, such as the attribute names available for that service. For example, some of the attribute names available for EC2 are volumeType, maxIopsVolume, operation, locationType, and instanceCapacity10xlarge.

This operation returns paginated results.

Creating a Request

data DescribeServices

See: newDescribeServices smart constructor.

Constructors

DescribeServices'

Fields

nextToken :: Maybe Text

The pagination token that indicates the next set of results that you want to retrieve.

maxResults :: Maybe Natural

The maximum number of results that you want returned in the response.

serviceCode :: Maybe Text

The code for the service whose information you want to retrieve, such as AmazonEC2. You can use the ServiceCode to filter the results in a GetProducts call. To retrieve a list of all services, leave this blank.

formatVersion :: Maybe Text

The format version that you want the response to be in.

Valid values are: aws_v1

newDescribeServices :: DescribeServices

Create a value of DescribeServices with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeServices', describeServices_nextToken - The pagination token that indicates the next set of results that you want to retrieve.

$sel:maxResults:DescribeServices', describeServices_maxResults - The maximum number of results that you want returned in the response.

$sel:serviceCode:DescribeServices', describeServices_serviceCode - The code for the service whose information you want to retrieve, such as AmazonEC2. You can use the ServiceCode to filter the results in a GetProducts call. To retrieve a list of all services, leave this blank.

$sel:formatVersion:DescribeServices', describeServices_formatVersion - The format version that you want the response to be in.

Valid values are: aws_v1

Request Lenses

describeServices_nextToken :: Lens' DescribeServices (Maybe Text)

The pagination token that indicates the next set of results that you want to retrieve.

describeServices_maxResults :: Lens' DescribeServices (Maybe Natural)

The maximum number of results that you want returned in the response.

describeServices_serviceCode :: Lens' DescribeServices (Maybe Text)

The code for the service whose information you want to retrieve, such as AmazonEC2. You can use the ServiceCode to filter the results in a GetProducts call. To retrieve a list of all services, leave this blank.

describeServices_formatVersion :: Lens' DescribeServices (Maybe Text)

The format version that you want the response to be in.

Valid values are: aws_v1

Destructuring the Response

data DescribeServicesResponse

See: newDescribeServicesResponse smart constructor.

Constructors

DescribeServicesResponse'

Fields

nextToken :: Maybe Text

The pagination token for the next set of retreivable results.

services :: Maybe [PricingService]

The service metadata for the service or services in the response.

formatVersion :: Maybe Text

The format version of the response. For example, aws_v1.

httpStatus :: Int

The response's http status code.

newDescribeServicesResponse

Arguments

:: Int
$sel:httpStatus:DescribeServicesResponse'

-> DescribeServicesResponse

Create a value of DescribeServicesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:DescribeServices', describeServicesResponse_nextToken - The pagination token for the next set of retreivable results.

$sel:services:DescribeServicesResponse', describeServicesResponse_services - The service metadata for the service or services in the response.

$sel:formatVersion:DescribeServices', describeServicesResponse_formatVersion - The format version of the response. For example, aws_v1.

$sel:httpStatus:DescribeServicesResponse', describeServicesResponse_httpStatus - The response's http status code.

Response Lenses

describeServicesResponse_nextToken :: Lens' DescribeServicesResponse (Maybe Text)

The pagination token for the next set of retreivable results.

describeServicesResponse_services :: Lens' DescribeServicesResponse (Maybe [PricingService])

The service metadata for the service or services in the response.

describeServicesResponse_formatVersion :: Lens' DescribeServicesResponse (Maybe Text)

The format version of the response. For example, aws_v1.

describeServicesResponse_httpStatus :: Lens' DescribeServicesResponse Int

The response's http status code.