Network.AWS.Pricing.GetProducts

Description

Returns a list of all products that match the filter criteria.

This operation returns paginated results.

Creating a Request

data GetProducts

See: newGetProducts smart constructor.

Constructors

GetProducts'

Fields

nextToken :: Maybe Text

The pagination token that indicates the next set of results that you want to retrieve.

maxResults :: Maybe Natural

The maximum number of results to return in the response.

serviceCode :: Maybe Text

The code for the service whose products you want to retrieve.

formatVersion :: Maybe Text

The format version that you want the response to be in.

Valid values are: aws_v1

filters :: Maybe [Filter]

The list of filters that limit the returned products. only products that match all filters are returned.

newGetProducts :: GetProducts

Create a value of GetProducts with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetProducts', getProducts_nextToken - The pagination token that indicates the next set of results that you want to retrieve.

$sel:maxResults:GetProducts', getProducts_maxResults - The maximum number of results to return in the response.

$sel:serviceCode:GetProducts', getProducts_serviceCode - The code for the service whose products you want to retrieve.

$sel:formatVersion:GetProducts', getProducts_formatVersion - The format version that you want the response to be in.

Valid values are: aws_v1

$sel:filters:GetProducts', getProducts_filters - The list of filters that limit the returned products. only products that match all filters are returned.

Request Lenses

getProducts_nextToken :: Lens' GetProducts (Maybe Text)

The pagination token that indicates the next set of results that you want to retrieve.

getProducts_maxResults :: Lens' GetProducts (Maybe Natural)

The maximum number of results to return in the response.

getProducts_serviceCode :: Lens' GetProducts (Maybe Text)

The code for the service whose products you want to retrieve.

getProducts_formatVersion :: Lens' GetProducts (Maybe Text)

The format version that you want the response to be in.

Valid values are: aws_v1

getProducts_filters :: Lens' GetProducts (Maybe [Filter])

The list of filters that limit the returned products. only products that match all filters are returned.

Destructuring the Response

data GetProductsResponse

See: newGetProductsResponse smart constructor.

Constructors

GetProductsResponse'

Fields

priceList :: Maybe [Text]

The list of products that match your filters. The list contains both the product metadata and the price information.

nextToken :: Maybe Text

The pagination token that indicates the next set of results to retrieve.

formatVersion :: Maybe Text

The format version of the response. For example, aws_v1.

httpStatus :: Int

The response's http status code.

newGetProductsResponse

Arguments

:: Int
$sel:httpStatus:GetProductsResponse'

-> GetProductsResponse

Create a value of GetProductsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:priceList:GetProductsResponse', getProductsResponse_priceList - The list of products that match your filters. The list contains both the product metadata and the price information.

$sel:nextToken:GetProducts', getProductsResponse_nextToken - The pagination token that indicates the next set of results to retrieve.

$sel:formatVersion:GetProducts', getProductsResponse_formatVersion - The format version of the response. For example, aws_v1.

$sel:httpStatus:GetProductsResponse', getProductsResponse_httpStatus - The response's http status code.

Response Lenses

getProductsResponse_priceList :: Lens' GetProductsResponse (Maybe [Text])

The list of products that match your filters. The list contains both the product metadata and the price information.

getProductsResponse_nextToken :: Lens' GetProductsResponse (Maybe Text)

The pagination token that indicates the next set of results to retrieve.

getProductsResponse_formatVersion :: Lens' GetProductsResponse (Maybe Text)

The format version of the response. For example, aws_v1.

getProductsResponse_httpStatus :: Lens' GetProductsResponse Int

The response's http status code.