Network.AWS.Pricing.Types

Description

Service Configuration

defaultService :: Service

API version 2017-10-15 of the Amazon Price List Service SDK configuration.

Errors

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

One or more parameters had an invalid value.

_InternalErrorException :: AsError a => Getting (First ServiceError) a ServiceError

An error on the server occurred during the processing of your request. Try again later.

_ExpiredNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The pagination token expired. Try again without a pagination token.

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The requested resource can't be found.

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The pagination token is invalid. Try again without a pagination token.

FilterType

newtype FilterType

Constructors

FilterType'

Fields

fromFilterType :: Text

Bundled Patterns

pattern FilterType_TERM_MATCH :: FilterType

AttributeValue

data AttributeValue

The values of a given attribute, such as Throughput Optimized HDD or Provisioned IOPS for the Amazon EC2 volumeType attribute.

See: newAttributeValue smart constructor.

Constructors

AttributeValue'

Fields

value :: Maybe Text

The specific value of an attributeName.

newAttributeValue :: AttributeValue

Create a value of AttributeValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:AttributeValue', attributeValue_value - The specific value of an attributeName.

attributeValue_value :: Lens' AttributeValue (Maybe Text)

The specific value of an attributeName.

Filter

data Filter

The constraints that you want all returned products to match.

See: newFilter smart constructor.

Constructors

Filter'

Fields

type' :: FilterType

The type of filter that you want to use.

Valid values are: TERM_MATCH. TERM_MATCH returns only products that match both the given filter field and the given value.

field :: Text

The product metadata field that you want to filter on. You can filter by just the service code to see all products for a specific service, filter by just the attribute name to see a specific attribute for multiple services, or use both a service code and an attribute name to retrieve only products that match both fields.

Valid values include: ServiceCode, and all attribute names

For example, you can filter by the AmazonEC2 service code and the volumeType attribute name to get the prices for only Amazon EC2 volumes.

value :: Text

The service code or attribute value that you want to filter by. If you are filtering by service code this is the actual service code, such as AmazonEC2. If you are filtering by attribute name, this is the attribute value that you want the returned products to match, such as a Provisioned IOPS volume.

newFilter

Arguments

:: FilterType
$sel:type':Filter'

-> Text
$sel:field:Filter'

-> Text
$sel:value:Filter'

-> Filter

Create a value of Filter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:type':Filter', filter_type - The type of filter that you want to use.

Valid values are: TERM_MATCH. TERM_MATCH returns only products that match both the given filter field and the given value.

$sel:field:Filter', filter_field - The product metadata field that you want to filter on. You can filter by just the service code to see all products for a specific service, filter by just the attribute name to see a specific attribute for multiple services, or use both a service code and an attribute name to retrieve only products that match both fields.

Valid values include: ServiceCode, and all attribute names

For example, you can filter by the AmazonEC2 service code and the volumeType attribute name to get the prices for only Amazon EC2 volumes.

$sel:value:Filter', filter_value - The service code or attribute value that you want to filter by. If you are filtering by service code this is the actual service code, such as AmazonEC2. If you are filtering by attribute name, this is the attribute value that you want the returned products to match, such as a Provisioned IOPS volume.

filter_type :: Lens' Filter FilterType

The type of filter that you want to use.

Valid values are: TERM_MATCH. TERM_MATCH returns only products that match both the given filter field and the given value.

filter_field :: Lens' Filter Text

The product metadata field that you want to filter on. You can filter by just the service code to see all products for a specific service, filter by just the attribute name to see a specific attribute for multiple services, or use both a service code and an attribute name to retrieve only products that match both fields.

Valid values include: ServiceCode, and all attribute names

For example, you can filter by the AmazonEC2 service code and the volumeType attribute name to get the prices for only Amazon EC2 volumes.

filter_value :: Lens' Filter Text

The service code or attribute value that you want to filter by. If you are filtering by service code this is the actual service code, such as AmazonEC2. If you are filtering by attribute name, this is the attribute value that you want the returned products to match, such as a Provisioned IOPS volume.

PricingService

data PricingService

The metadata for a service, such as the service code and available attribute names.

See: newPricingService smart constructor.

Constructors

PricingService'

Fields

attributeNames :: Maybe [Text]

The attributes that are available for this service.

serviceCode :: Maybe Text

The code for the Amazon Web Services service.

newPricingService :: PricingService

Create a value of PricingService with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeNames:PricingService', pricingService_attributeNames - The attributes that are available for this service.

$sel:serviceCode:PricingService', pricingService_serviceCode - The code for the Amazon Web Services service.

pricingService_attributeNames :: Lens' PricingService (Maybe [Text])

The attributes that are available for this service.

pricingService_serviceCode :: Lens' PricingService (Maybe Text)

The code for the Amazon Web Services service.