Network.AWS.Pricing.Types.AttributeValue

Description

Documentation

data AttributeValue

The values of a given attribute, such as Throughput Optimized HDD or Provisioned IOPS for the Amazon EC2 volumeType attribute.

See: newAttributeValue smart constructor.

Constructors

AttributeValue'

Fields

value :: Maybe Text

The specific value of an attributeName.

newAttributeValue :: AttributeValue

Create a value of AttributeValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:AttributeValue', attributeValue_value - The specific value of an attributeName.

attributeValue_value :: Lens' AttributeValue (Maybe Text)

The specific value of an attributeName.