Network.AWS.Pricing.Types.PricingService

Description

Documentation

data PricingService

The metadata for a service, such as the service code and available attribute names.

See: newPricingService smart constructor.

Constructors

PricingService'

Fields

serviceCode :: Maybe Text

The code for the Amazon Web Services service.

attributeNames :: Maybe [Text]

The attributes that are available for this service.

newPricingService :: PricingService

Create a value of PricingService with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:serviceCode:PricingService', pricingService_serviceCode - The code for the Amazon Web Services service.

$sel:attributeNames:PricingService', pricingService_attributeNames - The attributes that are available for this service.

pricingService_serviceCode :: Lens' PricingService (Maybe Text)

The code for the Amazon Web Services service.

pricingService_attributeNames :: Lens' PricingService (Maybe [Text])

The attributes that are available for this service.