Network.AWS.Proton.GetServiceInstance

Description

Get detail data for a service instance. A service instance is an instantiation of service template, which is running in a specific environment.

Creating a Request

data GetServiceInstance

See: newGetServiceInstance smart constructor.

Constructors

GetServiceInstance'

Fields

name :: Text

The name of a service instance that you want to get the detail data for.

serviceName :: Text

The name of the service that the service instance belongs to.

newGetServiceInstance

Arguments

:: Text
$sel:name:GetServiceInstance'

-> Text
$sel:serviceName:GetServiceInstance'

-> GetServiceInstance

Create a value of GetServiceInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:GetServiceInstance', getServiceInstance_name - The name of a service instance that you want to get the detail data for.

$sel:serviceName:GetServiceInstance', getServiceInstance_serviceName - The name of the service that the service instance belongs to.

Request Lenses

getServiceInstance_name :: Lens' GetServiceInstance Text

The name of a service instance that you want to get the detail data for.

getServiceInstance_serviceName :: Lens' GetServiceInstance Text

The name of the service that the service instance belongs to.

Destructuring the Response

data GetServiceInstanceResponse

See: newGetServiceInstanceResponse smart constructor.

Constructors

GetServiceInstanceResponse'

Fields

httpStatus :: Int

The response's http status code.

serviceInstance :: ServiceInstance

The service instance detail data that's returned by AWS Proton.

newGetServiceInstanceResponse

Arguments

:: Int
$sel:httpStatus:GetServiceInstanceResponse'

-> ServiceInstance
$sel:serviceInstance:GetServiceInstanceResponse'

-> GetServiceInstanceResponse

Create a value of GetServiceInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:GetServiceInstanceResponse', getServiceInstanceResponse_httpStatus - The response's http status code.

$sel:serviceInstance:GetServiceInstanceResponse', getServiceInstanceResponse_serviceInstance - The service instance detail data that's returned by AWS Proton.

Response Lenses

getServiceInstanceResponse_httpStatus :: Lens' GetServiceInstanceResponse Int

The response's http status code.

getServiceInstanceResponse_serviceInstance :: Lens' GetServiceInstanceResponse ServiceInstance

The service instance detail data that's returned by AWS Proton.