Network.AWS.Proton.ListEnvironmentTemplateVersions

Description

List major or minor versions of an environment template with detail data.

This operation returns paginated results.

Creating a Request

data ListEnvironmentTemplateVersions

See: newListEnvironmentTemplateVersions smart constructor.

Constructors

ListEnvironmentTemplateVersions'

Fields

majorVersion :: Maybe Text

To view a list of minor of versions under a major version of an environment template, include majorVersion.

To view a list of major versions of an environment template, exclude majorVersion.

nextToken :: Maybe Text

A token to indicate the location of the next major or minor version in the array of major or minor versions of an environment template, after the list of major or minor versions that was previously requested.

maxResults :: Maybe Natural

The maximum number of major or minor versions of an environment template to list.

templateName :: Text

The name of the environment template.

newListEnvironmentTemplateVersions

Arguments

:: Text
$sel:templateName:ListEnvironmentTemplateVersions'

-> ListEnvironmentTemplateVersions

Create a value of ListEnvironmentTemplateVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:majorVersion:ListEnvironmentTemplateVersions', listEnvironmentTemplateVersions_majorVersion - To view a list of minor of versions under a major version of an environment template, include majorVersion.

To view a list of major versions of an environment template, exclude majorVersion.

$sel:nextToken:ListEnvironmentTemplateVersions', listEnvironmentTemplateVersions_nextToken - A token to indicate the location of the next major or minor version in the array of major or minor versions of an environment template, after the list of major or minor versions that was previously requested.

$sel:maxResults:ListEnvironmentTemplateVersions', listEnvironmentTemplateVersions_maxResults - The maximum number of major or minor versions of an environment template to list.

$sel:templateName:ListEnvironmentTemplateVersions', listEnvironmentTemplateVersions_templateName - The name of the environment template.

Request Lenses

listEnvironmentTemplateVersions_majorVersion :: Lens' ListEnvironmentTemplateVersions (Maybe Text)

To view a list of minor of versions under a major version of an environment template, include majorVersion.

To view a list of major versions of an environment template, exclude majorVersion.

listEnvironmentTemplateVersions_nextToken :: Lens' ListEnvironmentTemplateVersions (Maybe Text)

A token to indicate the location of the next major or minor version in the array of major or minor versions of an environment template, after the list of major or minor versions that was previously requested.

listEnvironmentTemplateVersions_maxResults :: Lens' ListEnvironmentTemplateVersions (Maybe Natural)

The maximum number of major or minor versions of an environment template to list.

listEnvironmentTemplateVersions_templateName :: Lens' ListEnvironmentTemplateVersions Text

The name of the environment template.

Destructuring the Response

data ListEnvironmentTemplateVersionsResponse

See: newListEnvironmentTemplateVersionsResponse smart constructor.

Constructors

ListEnvironmentTemplateVersionsResponse'

Fields

nextToken :: Maybe Text

A token to indicate the location of the next major or minor version in the array of major or minor versions of an environment template, after the list of major or minor versions that was previously requested.

httpStatus :: Int

The response's http status code.

templateVersions :: [EnvironmentTemplateVersionSummary]

An array of major or minor versions of an environment template detail data.

newListEnvironmentTemplateVersionsResponse

Arguments

:: Int
$sel:httpStatus:ListEnvironmentTemplateVersionsResponse'

-> ListEnvironmentTemplateVersionsResponse

Create a value of ListEnvironmentTemplateVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEnvironmentTemplateVersions', listEnvironmentTemplateVersionsResponse_nextToken - A token to indicate the location of the next major or minor version in the array of major or minor versions of an environment template, after the list of major or minor versions that was previously requested.

$sel:httpStatus:ListEnvironmentTemplateVersionsResponse', listEnvironmentTemplateVersionsResponse_httpStatus - The response's http status code.

$sel:templateVersions:ListEnvironmentTemplateVersionsResponse', listEnvironmentTemplateVersionsResponse_templateVersions - An array of major or minor versions of an environment template detail data.

Response Lenses

listEnvironmentTemplateVersionsResponse_nextToken :: Lens' ListEnvironmentTemplateVersionsResponse (Maybe Text)

A token to indicate the location of the next major or minor version in the array of major or minor versions of an environment template, after the list of major or minor versions that was previously requested.

listEnvironmentTemplateVersionsResponse_httpStatus :: Lens' ListEnvironmentTemplateVersionsResponse Int

The response's http status code.

listEnvironmentTemplateVersionsResponse_templateVersions :: Lens' ListEnvironmentTemplateVersionsResponse [EnvironmentTemplateVersionSummary]

An array of major or minor versions of an environment template detail data.