Network.AWS.Proton.ListEnvironmentTemplates

Description

List environment templates.

This operation returns paginated results.

Creating a Request

data ListEnvironmentTemplates

See: newListEnvironmentTemplates smart constructor.

Constructors

ListEnvironmentTemplates'

Fields

nextToken :: Maybe Text

A token to indicate the location of the next environment template in the array of environment templates, after the list of environment templates that was previously requested.

maxResults :: Maybe Natural

The maximum number of environment templates to list.

newListEnvironmentTemplates :: ListEnvironmentTemplates

Create a value of ListEnvironmentTemplates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEnvironmentTemplates', listEnvironmentTemplates_nextToken - A token to indicate the location of the next environment template in the array of environment templates, after the list of environment templates that was previously requested.

$sel:maxResults:ListEnvironmentTemplates', listEnvironmentTemplates_maxResults - The maximum number of environment templates to list.

Request Lenses

listEnvironmentTemplates_nextToken :: Lens' ListEnvironmentTemplates (Maybe Text)

A token to indicate the location of the next environment template in the array of environment templates, after the list of environment templates that was previously requested.

listEnvironmentTemplates_maxResults :: Lens' ListEnvironmentTemplates (Maybe Natural)

The maximum number of environment templates to list.

Destructuring the Response

data ListEnvironmentTemplatesResponse

See: newListEnvironmentTemplatesResponse smart constructor.

Constructors

ListEnvironmentTemplatesResponse'

Fields

nextToken :: Maybe Text

A token to indicate the location of the next environment template in the array of environment templates, after the current requested list of environment templates.

httpStatus :: Int

The response's http status code.

templates :: [EnvironmentTemplateSummary]

An array of environment templates with detail data.

newListEnvironmentTemplatesResponse

Arguments

:: Int
$sel:httpStatus:ListEnvironmentTemplatesResponse'

-> ListEnvironmentTemplatesResponse

Create a value of ListEnvironmentTemplatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListEnvironmentTemplates', listEnvironmentTemplatesResponse_nextToken - A token to indicate the location of the next environment template in the array of environment templates, after the current requested list of environment templates.

$sel:httpStatus:ListEnvironmentTemplatesResponse', listEnvironmentTemplatesResponse_httpStatus - The response's http status code.

$sel:templates:ListEnvironmentTemplatesResponse', listEnvironmentTemplatesResponse_templates - An array of environment templates with detail data.

Response Lenses

listEnvironmentTemplatesResponse_nextToken :: Lens' ListEnvironmentTemplatesResponse (Maybe Text)

A token to indicate the location of the next environment template in the array of environment templates, after the current requested list of environment templates.

listEnvironmentTemplatesResponse_httpStatus :: Lens' ListEnvironmentTemplatesResponse Int

The response's http status code.

listEnvironmentTemplatesResponse_templates :: Lens' ListEnvironmentTemplatesResponse [EnvironmentTemplateSummary]

An array of environment templates with detail data.