Network.AWS.Proton.ListServiceTemplateVersions

Description

List major or minor versions of a service template with detail data.

This operation returns paginated results.

Creating a Request

data ListServiceTemplateVersions

See: newListServiceTemplateVersions smart constructor.

Constructors

ListServiceTemplateVersions'

Fields

majorVersion :: Maybe Text

To view a list of minor of versions under a major version of a service template, include majorVersion.

To view a list of major versions of a service template, exclude majorVersion.

nextToken :: Maybe Text

A token to indicate the location of the next major or minor version in the array of major or minor versions of a service template, after the list of major or minor versions that was previously requested.

maxResults :: Maybe Natural

The maximum number of major or minor versions of a service template to list.

templateName :: Text

The name of the service template.

newListServiceTemplateVersions

Arguments

:: Text
$sel:templateName:ListServiceTemplateVersions'

-> ListServiceTemplateVersions

Create a value of ListServiceTemplateVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:majorVersion:ListServiceTemplateVersions', listServiceTemplateVersions_majorVersion - To view a list of minor of versions under a major version of a service template, include majorVersion.

To view a list of major versions of a service template, exclude majorVersion.

$sel:nextToken:ListServiceTemplateVersions', listServiceTemplateVersions_nextToken - A token to indicate the location of the next major or minor version in the array of major or minor versions of a service template, after the list of major or minor versions that was previously requested.

$sel:maxResults:ListServiceTemplateVersions', listServiceTemplateVersions_maxResults - The maximum number of major or minor versions of a service template to list.

$sel:templateName:ListServiceTemplateVersions', listServiceTemplateVersions_templateName - The name of the service template.

Request Lenses

listServiceTemplateVersions_majorVersion :: Lens' ListServiceTemplateVersions (Maybe Text)

To view a list of minor of versions under a major version of a service template, include majorVersion.

To view a list of major versions of a service template, exclude majorVersion.

listServiceTemplateVersions_nextToken :: Lens' ListServiceTemplateVersions (Maybe Text)

A token to indicate the location of the next major or minor version in the array of major or minor versions of a service template, after the list of major or minor versions that was previously requested.

listServiceTemplateVersions_maxResults :: Lens' ListServiceTemplateVersions (Maybe Natural)

The maximum number of major or minor versions of a service template to list.

listServiceTemplateVersions_templateName :: Lens' ListServiceTemplateVersions Text

The name of the service template.

Destructuring the Response

data ListServiceTemplateVersionsResponse

See: newListServiceTemplateVersionsResponse smart constructor.

Constructors

ListServiceTemplateVersionsResponse'

Fields

nextToken :: Maybe Text

A token to indicate the location of the next major or minor version in the array of major or minor versions of a service template, after the list of major or minor versions that was previously requested.

httpStatus :: Int

The response's http status code.

templateVersions :: [ServiceTemplateVersionSummary]

An array of major or minor versions of a service template with detail data.

newListServiceTemplateVersionsResponse

Arguments

:: Int
$sel:httpStatus:ListServiceTemplateVersionsResponse'

-> ListServiceTemplateVersionsResponse

Create a value of ListServiceTemplateVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListServiceTemplateVersions', listServiceTemplateVersionsResponse_nextToken - A token to indicate the location of the next major or minor version in the array of major or minor versions of a service template, after the list of major or minor versions that was previously requested.

$sel:httpStatus:ListServiceTemplateVersionsResponse', listServiceTemplateVersionsResponse_httpStatus - The response's http status code.

$sel:templateVersions:ListServiceTemplateVersionsResponse', listServiceTemplateVersionsResponse_templateVersions - An array of major or minor versions of a service template with detail data.

Response Lenses

listServiceTemplateVersionsResponse_nextToken :: Lens' ListServiceTemplateVersionsResponse (Maybe Text)

A token to indicate the location of the next major or minor version in the array of major or minor versions of a service template, after the list of major or minor versions that was previously requested.

listServiceTemplateVersionsResponse_httpStatus :: Lens' ListServiceTemplateVersionsResponse Int

The response's http status code.

listServiceTemplateVersionsResponse_templateVersions :: Lens' ListServiceTemplateVersionsResponse [ServiceTemplateVersionSummary]

An array of major or minor versions of a service template with detail data.