Network.AWS.Proton.Types.EnvironmentTemplate

Description

Documentation

data EnvironmentTemplate

The environment template data.

See: newEnvironmentTemplate smart constructor.

Constructors

EnvironmentTemplate'

Fields

provisioning :: Maybe Provisioning

When included, indicates that the environment template is for customer provisioned and managed infrastructure.

recommendedVersion :: Maybe Text

The ID of the recommended version of the environment template.

displayName :: Maybe (Sensitive Text)

The name of the environment template as displayed in the developer interface.

encryptionKey :: Maybe Text

The customer provided encryption key for the environment template.

description :: Maybe (Sensitive Text)

A description of the environment template.

arn :: Text

The Amazon Resource Name (ARN) of the environment template.

createdAt :: POSIX

The time when the environment template was created.

lastModifiedAt :: POSIX

The time when the environment template was last modified.

name :: Text

The name of the environment template.

newEnvironmentTemplate

Arguments

:: Text
$sel:arn:EnvironmentTemplate'

-> UTCTime
$sel:createdAt:EnvironmentTemplate'

-> UTCTime
$sel:lastModifiedAt:EnvironmentTemplate'

-> Text
$sel:name:EnvironmentTemplate'

-> EnvironmentTemplate

Create a value of EnvironmentTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:provisioning:EnvironmentTemplate', environmentTemplate_provisioning - When included, indicates that the environment template is for customer provisioned and managed infrastructure.

$sel:recommendedVersion:EnvironmentTemplate', environmentTemplate_recommendedVersion - The ID of the recommended version of the environment template.

$sel:displayName:EnvironmentTemplate', environmentTemplate_displayName - The name of the environment template as displayed in the developer interface.

$sel:encryptionKey:EnvironmentTemplate', environmentTemplate_encryptionKey - The customer provided encryption key for the environment template.

$sel:description:EnvironmentTemplate', environmentTemplate_description - A description of the environment template.

$sel:arn:EnvironmentTemplate', environmentTemplate_arn - The Amazon Resource Name (ARN) of the environment template.

$sel:createdAt:EnvironmentTemplate', environmentTemplate_createdAt - The time when the environment template was created.

$sel:lastModifiedAt:EnvironmentTemplate', environmentTemplate_lastModifiedAt - The time when the environment template was last modified.

$sel:name:EnvironmentTemplate', environmentTemplate_name - The name of the environment template.

environmentTemplate_provisioning :: Lens' EnvironmentTemplate (Maybe Provisioning)

When included, indicates that the environment template is for customer provisioned and managed infrastructure.

environmentTemplate_recommendedVersion :: Lens' EnvironmentTemplate (Maybe Text)

The ID of the recommended version of the environment template.

environmentTemplate_displayName :: Lens' EnvironmentTemplate (Maybe Text)

The name of the environment template as displayed in the developer interface.

environmentTemplate_encryptionKey :: Lens' EnvironmentTemplate (Maybe Text)

The customer provided encryption key for the environment template.

environmentTemplate_description :: Lens' EnvironmentTemplate (Maybe Text)

A description of the environment template.

environmentTemplate_arn :: Lens' EnvironmentTemplate Text

The Amazon Resource Name (ARN) of the environment template.

environmentTemplate_createdAt :: Lens' EnvironmentTemplate UTCTime

The time when the environment template was created.

environmentTemplate_lastModifiedAt :: Lens' EnvironmentTemplate UTCTime

The time when the environment template was last modified.

environmentTemplate_name :: Lens' EnvironmentTemplate Text

The name of the environment template.