Network.AWS.Proton.Types.ServiceInstance

Description

Documentation

data ServiceInstance

The service instance detail data.

See: newServiceInstance smart constructor.

Constructors

ServiceInstance'

Fields

deploymentStatusMessage :: Maybe (Sensitive Text)

A service instance deployment status message.

spec :: Maybe (Sensitive Text)

The service spec that was used to create the service instance.

arn :: Text

The Amazon Resource Name (ARN) of the service instance.

createdAt :: POSIX

The time when the service instance was created.

deploymentStatus :: DeploymentStatus

The service instance deployment status.

environmentName :: Text

The name of the environment that the service instance was deployed into.

lastDeploymentAttemptedAt :: POSIX

The time when a deployment of the service instance was last attempted.

lastDeploymentSucceededAt :: POSIX

The time when the service instance was last deployed successfully.

name :: Text

The name of the service instance.

serviceName :: Text

The name of the service that the service instance belongs to.

templateMajorVersion :: Text

The ID of the major version of the service template that was used to create the service instance.

templateMinorVersion :: Text

The ID of the minor version of the service template that was used to create the service instance.

templateName :: Text

The name of the service template that was used to create the service instance.

newServiceInstance

Arguments

:: Text
$sel:arn:ServiceInstance'

-> UTCTime
$sel:createdAt:ServiceInstance'

-> DeploymentStatus
$sel:deploymentStatus:ServiceInstance'

-> Text
$sel:environmentName:ServiceInstance'

-> UTCTime
$sel:lastDeploymentAttemptedAt:ServiceInstance'

-> UTCTime
$sel:lastDeploymentSucceededAt:ServiceInstance'

-> Text
$sel:name:ServiceInstance'

-> Text
$sel:serviceName:ServiceInstance'

-> Text
$sel:templateMajorVersion:ServiceInstance'

-> Text
$sel:templateMinorVersion:ServiceInstance'

-> Text
$sel:templateName:ServiceInstance'

-> ServiceInstance

Create a value of ServiceInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentStatusMessage:ServiceInstance', serviceInstance_deploymentStatusMessage - A service instance deployment status message.

$sel:spec:ServiceInstance', serviceInstance_spec - The service spec that was used to create the service instance.

$sel:arn:ServiceInstance', serviceInstance_arn - The Amazon Resource Name (ARN) of the service instance.

$sel:createdAt:ServiceInstance', serviceInstance_createdAt - The time when the service instance was created.

$sel:deploymentStatus:ServiceInstance', serviceInstance_deploymentStatus - The service instance deployment status.

$sel:environmentName:ServiceInstance', serviceInstance_environmentName - The name of the environment that the service instance was deployed into.

$sel:lastDeploymentAttemptedAt:ServiceInstance', serviceInstance_lastDeploymentAttemptedAt - The time when a deployment of the service instance was last attempted.

$sel:lastDeploymentSucceededAt:ServiceInstance', serviceInstance_lastDeploymentSucceededAt - The time when the service instance was last deployed successfully.

$sel:name:ServiceInstance', serviceInstance_name - The name of the service instance.

$sel:serviceName:ServiceInstance', serviceInstance_serviceName - The name of the service that the service instance belongs to.

$sel:templateMajorVersion:ServiceInstance', serviceInstance_templateMajorVersion - The ID of the major version of the service template that was used to create the service instance.

$sel:templateMinorVersion:ServiceInstance', serviceInstance_templateMinorVersion - The ID of the minor version of the service template that was used to create the service instance.

$sel:templateName:ServiceInstance', serviceInstance_templateName - The name of the service template that was used to create the service instance.

serviceInstance_deploymentStatusMessage :: Lens' ServiceInstance (Maybe Text)

A service instance deployment status message.

serviceInstance_spec :: Lens' ServiceInstance (Maybe Text)

The service spec that was used to create the service instance.

serviceInstance_arn :: Lens' ServiceInstance Text

The Amazon Resource Name (ARN) of the service instance.

serviceInstance_createdAt :: Lens' ServiceInstance UTCTime

The time when the service instance was created.

serviceInstance_deploymentStatus :: Lens' ServiceInstance DeploymentStatus

The service instance deployment status.

serviceInstance_environmentName :: Lens' ServiceInstance Text

The name of the environment that the service instance was deployed into.

serviceInstance_lastDeploymentAttemptedAt :: Lens' ServiceInstance UTCTime

The time when a deployment of the service instance was last attempted.

serviceInstance_lastDeploymentSucceededAt :: Lens' ServiceInstance UTCTime

The time when the service instance was last deployed successfully.

serviceInstance_name :: Lens' ServiceInstance Text

The name of the service instance.

serviceInstance_serviceName :: Lens' ServiceInstance Text

The name of the service that the service instance belongs to.

serviceInstance_templateMajorVersion :: Lens' ServiceInstance Text

The ID of the major version of the service template that was used to create the service instance.

serviceInstance_templateMinorVersion :: Lens' ServiceInstance Text

The ID of the minor version of the service template that was used to create the service instance.

serviceInstance_templateName :: Lens' ServiceInstance Text

The name of the service template that was used to create the service instance.