Network.AWS.Proton.Types.ServicePipeline

Description

Documentation

data ServicePipeline

The service pipeline detail data.

See: newServicePipeline smart constructor.

Constructors

ServicePipeline'

Fields

deploymentStatusMessage :: Maybe (Sensitive Text)

A service pipeline deployment status message.

spec :: Maybe (Sensitive Text)

The service spec that was used to create the service pipeline.

arn :: Text

The Amazon Resource Name (ARN) of the service pipeline.

createdAt :: POSIX

The time when the service pipeline was created.

deploymentStatus :: DeploymentStatus

The deployment status of the service pipeline.

lastDeploymentAttemptedAt :: POSIX

The time when a deployment of the service pipeline was last attempted.

lastDeploymentSucceededAt :: POSIX

The time when the service pipeline was last deployed successfully.

templateMajorVersion :: Text

The ID of the major version of the service template that was used to create the service pipeline.

templateMinorVersion :: Text

The ID of the minor version of the service template that was used to create the service pipeline.

templateName :: Text

The name of the service template that was used to create the service pipeline.

newServicePipeline

Arguments

:: Text
$sel:arn:ServicePipeline'

-> UTCTime
$sel:createdAt:ServicePipeline'

-> DeploymentStatus
$sel:deploymentStatus:ServicePipeline'

-> UTCTime
$sel:lastDeploymentAttemptedAt:ServicePipeline'

-> UTCTime
$sel:lastDeploymentSucceededAt:ServicePipeline'

-> Text
$sel:templateMajorVersion:ServicePipeline'

-> Text
$sel:templateMinorVersion:ServicePipeline'

-> Text
$sel:templateName:ServicePipeline'

-> ServicePipeline

Create a value of ServicePipeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deploymentStatusMessage:ServicePipeline', servicePipeline_deploymentStatusMessage - A service pipeline deployment status message.

$sel:spec:ServicePipeline', servicePipeline_spec - The service spec that was used to create the service pipeline.

$sel:arn:ServicePipeline', servicePipeline_arn - The Amazon Resource Name (ARN) of the service pipeline.

$sel:createdAt:ServicePipeline', servicePipeline_createdAt - The time when the service pipeline was created.

$sel:deploymentStatus:ServicePipeline', servicePipeline_deploymentStatus - The deployment status of the service pipeline.

$sel:lastDeploymentAttemptedAt:ServicePipeline', servicePipeline_lastDeploymentAttemptedAt - The time when a deployment of the service pipeline was last attempted.

$sel:lastDeploymentSucceededAt:ServicePipeline', servicePipeline_lastDeploymentSucceededAt - The time when the service pipeline was last deployed successfully.

$sel:templateMajorVersion:ServicePipeline', servicePipeline_templateMajorVersion - The ID of the major version of the service template that was used to create the service pipeline.

$sel:templateMinorVersion:ServicePipeline', servicePipeline_templateMinorVersion - The ID of the minor version of the service template that was used to create the service pipeline.

$sel:templateName:ServicePipeline', servicePipeline_templateName - The name of the service template that was used to create the service pipeline.

servicePipeline_deploymentStatusMessage :: Lens' ServicePipeline (Maybe Text)

A service pipeline deployment status message.

servicePipeline_spec :: Lens' ServicePipeline (Maybe Text)

The service spec that was used to create the service pipeline.

servicePipeline_arn :: Lens' ServicePipeline Text

The Amazon Resource Name (ARN) of the service pipeline.

servicePipeline_createdAt :: Lens' ServicePipeline UTCTime

The time when the service pipeline was created.

servicePipeline_deploymentStatus :: Lens' ServicePipeline DeploymentStatus

The deployment status of the service pipeline.

servicePipeline_lastDeploymentAttemptedAt :: Lens' ServicePipeline UTCTime

The time when a deployment of the service pipeline was last attempted.

servicePipeline_lastDeploymentSucceededAt :: Lens' ServicePipeline UTCTime

The time when the service pipeline was last deployed successfully.

servicePipeline_templateMajorVersion :: Lens' ServicePipeline Text

The ID of the major version of the service template that was used to create the service pipeline.

servicePipeline_templateMinorVersion :: Lens' ServicePipeline Text

The ID of the minor version of the service template that was used to create the service pipeline.

servicePipeline_templateName :: Lens' ServicePipeline Text

The name of the service template that was used to create the service pipeline.