Network.AWS.Proton.Types.ServiceTemplateVersion

Description

Documentation

data ServiceTemplateVersion

The version of a service template detail data.

See: newServiceTemplateVersion smart constructor.

Constructors

ServiceTemplateVersion'

Fields

schema :: Maybe (Sensitive Text)

The schema of the version of a service template.

statusMessage :: Maybe (Sensitive Text)

A service template version status message.

recommendedMinorVersion :: Maybe Text

The ID of the recommended minor version of the service template.

description :: Maybe (Sensitive Text)

A description of the version of a service template.

arn :: Text

The Amazon Resource Name (ARN) of the version of a service template.

compatibleEnvironmentTemplates :: [CompatibleEnvironmentTemplate]

An array of compatible environment template names for the major version of a service template.

createdAt :: POSIX

The time when the version of a service template was created.

lastModifiedAt :: POSIX

The time when the version of a service template was last modified.

majorVersion :: Text

The ID of the latest major version that's associated with the version of a service template.

minorVersion :: Text

The ID of the minor version of a service template.

status :: TemplateVersionStatus

The service template version status.

templateName :: Text

The name of the version of a service template.

newServiceTemplateVersion

Arguments

:: Text
$sel:arn:ServiceTemplateVersion'

-> UTCTime
$sel:createdAt:ServiceTemplateVersion'

-> UTCTime
$sel:lastModifiedAt:ServiceTemplateVersion'

-> Text
$sel:majorVersion:ServiceTemplateVersion'

-> Text
$sel:minorVersion:ServiceTemplateVersion'

-> TemplateVersionStatus
$sel:status:ServiceTemplateVersion'

-> Text
$sel:templateName:ServiceTemplateVersion'

-> ServiceTemplateVersion

Create a value of ServiceTemplateVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:schema:ServiceTemplateVersion', serviceTemplateVersion_schema - The schema of the version of a service template.

$sel:statusMessage:ServiceTemplateVersion', serviceTemplateVersion_statusMessage - A service template version status message.

$sel:recommendedMinorVersion:ServiceTemplateVersion', serviceTemplateVersion_recommendedMinorVersion - The ID of the recommended minor version of the service template.

$sel:description:ServiceTemplateVersion', serviceTemplateVersion_description - A description of the version of a service template.

$sel:arn:ServiceTemplateVersion', serviceTemplateVersion_arn - The Amazon Resource Name (ARN) of the version of a service template.

$sel:compatibleEnvironmentTemplates:ServiceTemplateVersion', serviceTemplateVersion_compatibleEnvironmentTemplates - An array of compatible environment template names for the major version of a service template.

$sel:createdAt:ServiceTemplateVersion', serviceTemplateVersion_createdAt - The time when the version of a service template was created.

$sel:lastModifiedAt:ServiceTemplateVersion', serviceTemplateVersion_lastModifiedAt - The time when the version of a service template was last modified.

$sel:majorVersion:ServiceTemplateVersion', serviceTemplateVersion_majorVersion - The ID of the latest major version that's associated with the version of a service template.

$sel:minorVersion:ServiceTemplateVersion', serviceTemplateVersion_minorVersion - The ID of the minor version of a service template.

$sel:status:ServiceTemplateVersion', serviceTemplateVersion_status - The service template version status.

$sel:templateName:ServiceTemplateVersion', serviceTemplateVersion_templateName - The name of the version of a service template.

serviceTemplateVersion_schema :: Lens' ServiceTemplateVersion (Maybe Text)

The schema of the version of a service template.

serviceTemplateVersion_statusMessage :: Lens' ServiceTemplateVersion (Maybe Text)

A service template version status message.

serviceTemplateVersion_recommendedMinorVersion :: Lens' ServiceTemplateVersion (Maybe Text)

The ID of the recommended minor version of the service template.

serviceTemplateVersion_description :: Lens' ServiceTemplateVersion (Maybe Text)

A description of the version of a service template.

serviceTemplateVersion_arn :: Lens' ServiceTemplateVersion Text

The Amazon Resource Name (ARN) of the version of a service template.

serviceTemplateVersion_compatibleEnvironmentTemplates :: Lens' ServiceTemplateVersion [CompatibleEnvironmentTemplate]

An array of compatible environment template names for the major version of a service template.

serviceTemplateVersion_createdAt :: Lens' ServiceTemplateVersion UTCTime

The time when the version of a service template was created.

serviceTemplateVersion_lastModifiedAt :: Lens' ServiceTemplateVersion UTCTime

The time when the version of a service template was last modified.

serviceTemplateVersion_majorVersion :: Lens' ServiceTemplateVersion Text

The ID of the latest major version that's associated with the version of a service template.

serviceTemplateVersion_minorVersion :: Lens' ServiceTemplateVersion Text

The ID of the minor version of a service template.

serviceTemplateVersion_status :: Lens' ServiceTemplateVersion TemplateVersionStatus

The service template version status.

serviceTemplateVersion_templateName :: Lens' ServiceTemplateVersion Text

The name of the version of a service template.