Network.AWS.Proton.Types.ServiceTemplateVersionSummary

Description

Documentation

data ServiceTemplateVersionSummary

A summary of the service template version detail data.

See: newServiceTemplateVersionSummary smart constructor.

Constructors

ServiceTemplateVersionSummary'

Fields

statusMessage :: Maybe (Sensitive Text)

A service template minor version status message.

recommendedMinorVersion :: Maybe Text

The ID of the recommended minor version of the service template.

description :: Maybe (Sensitive Text)

A description of the version of a service template.

arn :: Text

The Amazon Resource Name (ARN) of the version of a service template.

createdAt :: POSIX

The time when the version of a service template was created.

lastModifiedAt :: POSIX

The time when the version of a service template was last modified.

majorVersion :: Text

The ID of the latest major version that's associated with the version of a service template.

minorVersion :: Text

The ID of the minor version of a service template.

status :: TemplateVersionStatus

The service template minor version status.

templateName :: Text

The name of the service template.

newServiceTemplateVersionSummary

Arguments

:: Text
$sel:arn:ServiceTemplateVersionSummary'

-> UTCTime
$sel:createdAt:ServiceTemplateVersionSummary'

-> UTCTime
$sel:lastModifiedAt:ServiceTemplateVersionSummary'

-> Text
$sel:majorVersion:ServiceTemplateVersionSummary'

-> Text
$sel:minorVersion:ServiceTemplateVersionSummary'

-> TemplateVersionStatus
$sel:status:ServiceTemplateVersionSummary'

-> Text
$sel:templateName:ServiceTemplateVersionSummary'

-> ServiceTemplateVersionSummary

Create a value of ServiceTemplateVersionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:statusMessage:ServiceTemplateVersionSummary', serviceTemplateVersionSummary_statusMessage - A service template minor version status message.

$sel:recommendedMinorVersion:ServiceTemplateVersionSummary', serviceTemplateVersionSummary_recommendedMinorVersion - The ID of the recommended minor version of the service template.

$sel:description:ServiceTemplateVersionSummary', serviceTemplateVersionSummary_description - A description of the version of a service template.

$sel:arn:ServiceTemplateVersionSummary', serviceTemplateVersionSummary_arn - The Amazon Resource Name (ARN) of the version of a service template.

$sel:createdAt:ServiceTemplateVersionSummary', serviceTemplateVersionSummary_createdAt - The time when the version of a service template was created.

$sel:lastModifiedAt:ServiceTemplateVersionSummary', serviceTemplateVersionSummary_lastModifiedAt - The time when the version of a service template was last modified.

$sel:majorVersion:ServiceTemplateVersionSummary', serviceTemplateVersionSummary_majorVersion - The ID of the latest major version that's associated with the version of a service template.

$sel:minorVersion:ServiceTemplateVersionSummary', serviceTemplateVersionSummary_minorVersion - The ID of the minor version of a service template.

$sel:status:ServiceTemplateVersionSummary', serviceTemplateVersionSummary_status - The service template minor version status.

$sel:templateName:ServiceTemplateVersionSummary', serviceTemplateVersionSummary_templateName - The name of the service template.

serviceTemplateVersionSummary_statusMessage :: Lens' ServiceTemplateVersionSummary (Maybe Text)

A service template minor version status message.

serviceTemplateVersionSummary_recommendedMinorVersion :: Lens' ServiceTemplateVersionSummary (Maybe Text)

The ID of the recommended minor version of the service template.

serviceTemplateVersionSummary_description :: Lens' ServiceTemplateVersionSummary (Maybe Text)

A description of the version of a service template.

serviceTemplateVersionSummary_arn :: Lens' ServiceTemplateVersionSummary Text

The Amazon Resource Name (ARN) of the version of a service template.

serviceTemplateVersionSummary_createdAt :: Lens' ServiceTemplateVersionSummary UTCTime

The time when the version of a service template was created.

serviceTemplateVersionSummary_lastModifiedAt :: Lens' ServiceTemplateVersionSummary UTCTime

The time when the version of a service template was last modified.

serviceTemplateVersionSummary_majorVersion :: Lens' ServiceTemplateVersionSummary Text

The ID of the latest major version that's associated with the version of a service template.

serviceTemplateVersionSummary_minorVersion :: Lens' ServiceTemplateVersionSummary Text

The ID of the minor version of a service template.

serviceTemplateVersionSummary_status :: Lens' ServiceTemplateVersionSummary TemplateVersionStatus

The service template minor version status.

serviceTemplateVersionSummary_templateName :: Lens' ServiceTemplateVersionSummary Text

The name of the service template.