Network.AWS.Proton.UpdateEnvironment

Description

Update an environment.

If the environment is associated with an environment account connection, don't update or include the protonServiceRoleArn parameter to update or connect to an environment account connection.

You can only update to a new environment account connection if it was created in the same environment account that the current environment account connection was created in and is associated with the current environment.

If the environment isn't associated with an environment account connection, don't update or include the environmentAccountConnectionId parameter to update or connect to an environment account connection.

You can update either the environmentAccountConnectionId or protonServiceRoleArn parameter and value. You can’t update both.

There are four modes for updating an environment as described in the following. The deploymentType field defines the mode.

[] NONE

In this mode, a deployment doesn't occur. Only the requested metadata parameters are updated.

[] CURRENT_VERSION

In this mode, the environment is deployed and updated with the new spec that you provide. Only requested parameters are updated. Don’t include minor or major version parameters when you use this deployment-type.

[] MINOR_VERSION

In this mode, the environment is deployed and updated with the published, recommended (latest) minor version of the current major version in use, by default. You can also specify a different minor version of the current major version in use.

[] MAJOR_VERSION

In this mode, the environment is deployed and updated with the published, recommended (latest) major and minor version of the current template, by default. You can also specify a different major version that's higher than the major version in use and a minor version (optional).

Creating a Request

data UpdateEnvironment

See: newUpdateEnvironment smart constructor.

Constructors

UpdateEnvironment'

Fields

protonServiceRoleArn :: Maybe Text

The Amazon Resource Name (ARN) of the AWS Proton service role that allows AWS Proton to make API calls to other services your behalf.

environmentAccountConnectionId :: Maybe Text

The ID of the environment account connection.

You can only update to a new environment account connection if it was created in the same environment account that the current environment account connection was created in and is associated with the current environment.

spec :: Maybe (Sensitive Text)

The formatted specification that defines the update.

templateMinorVersion :: Maybe Text

The ID of the minor version of the environment to update.

description :: Maybe (Sensitive Text)

A description of the environment update.

templateMajorVersion :: Maybe Text

The ID of the major version of the environment to update.

deploymentType :: DeploymentUpdateType

There are four modes for updating an environment as described in the following. The deploymentType field defines the mode.

[] NONE

In this mode, a deployment doesn't occur. Only the requested metadata parameters are updated.

[] CURRENT_VERSION

In this mode, the environment is deployed and updated with the new spec that you provide. Only requested parameters are updated. Don’t include minor or major version parameters when you use this deployment-type.

[] MINOR_VERSION

In this mode, the environment is deployed and updated with the published, recommended (latest) minor version of the current major version in use, by default. You can also specify a different minor version of the current major version in use.

[] MAJOR_VERSION

In this mode, the environment is deployed and updated with the published, recommended (latest) major and minor version of the current template, by default. You can also specify a different major version that is higher than the major version in use and a minor version (optional).

name :: Text

The name of the environment to update.

newUpdateEnvironment

Arguments

:: DeploymentUpdateType
$sel:deploymentType:UpdateEnvironment'

-> Text
$sel:name:UpdateEnvironment'

-> UpdateEnvironment

Create a value of UpdateEnvironment with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:protonServiceRoleArn:UpdateEnvironment', updateEnvironment_protonServiceRoleArn - The Amazon Resource Name (ARN) of the AWS Proton service role that allows AWS Proton to make API calls to other services your behalf.

$sel:environmentAccountConnectionId:UpdateEnvironment', updateEnvironment_environmentAccountConnectionId - The ID of the environment account connection.

You can only update to a new environment account connection if it was created in the same environment account that the current environment account connection was created in and is associated with the current environment.

$sel:spec:UpdateEnvironment', updateEnvironment_spec - The formatted specification that defines the update.

$sel:templateMinorVersion:UpdateEnvironment', updateEnvironment_templateMinorVersion - The ID of the minor version of the environment to update.

$sel:description:UpdateEnvironment', updateEnvironment_description - A description of the environment update.

$sel:templateMajorVersion:UpdateEnvironment', updateEnvironment_templateMajorVersion - The ID of the major version of the environment to update.

$sel:deploymentType:UpdateEnvironment', updateEnvironment_deploymentType - There are four modes for updating an environment as described in the following. The deploymentType field defines the mode.

[] NONE

In this mode, a deployment doesn't occur. Only the requested metadata parameters are updated.

[] CURRENT_VERSION

In this mode, the environment is deployed and updated with the new spec that you provide. Only requested parameters are updated. Don’t include minor or major version parameters when you use this deployment-type.

[] MINOR_VERSION

In this mode, the environment is deployed and updated with the published, recommended (latest) minor version of the current major version in use, by default. You can also specify a different minor version of the current major version in use.

[] MAJOR_VERSION

In this mode, the environment is deployed and updated with the published, recommended (latest) major and minor version of the current template, by default. You can also specify a different major version that is higher than the major version in use and a minor version (optional).

$sel:name:UpdateEnvironment', updateEnvironment_name - The name of the environment to update.

Request Lenses

updateEnvironment_protonServiceRoleArn :: Lens' UpdateEnvironment (Maybe Text)

The Amazon Resource Name (ARN) of the AWS Proton service role that allows AWS Proton to make API calls to other services your behalf.

updateEnvironment_environmentAccountConnectionId :: Lens' UpdateEnvironment (Maybe Text)

The ID of the environment account connection.

You can only update to a new environment account connection if it was created in the same environment account that the current environment account connection was created in and is associated with the current environment.

updateEnvironment_spec :: Lens' UpdateEnvironment (Maybe Text)

The formatted specification that defines the update.

updateEnvironment_templateMinorVersion :: Lens' UpdateEnvironment (Maybe Text)

The ID of the minor version of the environment to update.

updateEnvironment_description :: Lens' UpdateEnvironment (Maybe Text)

A description of the environment update.

updateEnvironment_templateMajorVersion :: Lens' UpdateEnvironment (Maybe Text)

The ID of the major version of the environment to update.

updateEnvironment_deploymentType :: Lens' UpdateEnvironment DeploymentUpdateType

There are four modes for updating an environment as described in the following. The deploymentType field defines the mode.

[] NONE

In this mode, a deployment doesn't occur. Only the requested metadata parameters are updated.

[] CURRENT_VERSION

In this mode, the environment is deployed and updated with the new spec that you provide. Only requested parameters are updated. Don’t include minor or major version parameters when you use this deployment-type.

[] MINOR_VERSION

In this mode, the environment is deployed and updated with the published, recommended (latest) minor version of the current major version in use, by default. You can also specify a different minor version of the current major version in use.

[] MAJOR_VERSION

In this mode, the environment is deployed and updated with the published, recommended (latest) major and minor version of the current template, by default. You can also specify a different major version that is higher than the major version in use and a minor version (optional).

updateEnvironment_name :: Lens' UpdateEnvironment Text

The name of the environment to update.

Destructuring the Response

data UpdateEnvironmentResponse

See: newUpdateEnvironmentResponse smart constructor.

Constructors

UpdateEnvironmentResponse'

Fields

httpStatus :: Int

The response's http status code.

environment :: Environment

The environment detail data that's returned by AWS Proton.

newUpdateEnvironmentResponse

Arguments

:: Int
$sel:httpStatus:UpdateEnvironmentResponse'

-> Environment
$sel:environment:UpdateEnvironmentResponse'

-> UpdateEnvironmentResponse

Create a value of UpdateEnvironmentResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateEnvironmentResponse', updateEnvironmentResponse_httpStatus - The response's http status code.

$sel:environment:UpdateEnvironmentResponse', updateEnvironmentResponse_environment - The environment detail data that's returned by AWS Proton.

Response Lenses

updateEnvironmentResponse_httpStatus :: Lens' UpdateEnvironmentResponse Int

The response's http status code.

updateEnvironmentResponse_environment :: Lens' UpdateEnvironmentResponse Environment

The environment detail data that's returned by AWS Proton.