Network.AWS.Proton.UpdateEnvironmentTemplate

Description

Update an environment template.

Creating a Request

data UpdateEnvironmentTemplate

See: newUpdateEnvironmentTemplate smart constructor.

Constructors

UpdateEnvironmentTemplate'

Fields

displayName :: Maybe (Sensitive Text)

The name of the environment template to update as displayed in the developer interface.

description :: Maybe (Sensitive Text)

A description of the environment template update.

name :: Text

The name of the environment template to update.

newUpdateEnvironmentTemplate

Arguments

:: Text
$sel:name:UpdateEnvironmentTemplate'

-> UpdateEnvironmentTemplate

Create a value of UpdateEnvironmentTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:displayName:UpdateEnvironmentTemplate', updateEnvironmentTemplate_displayName - The name of the environment template to update as displayed in the developer interface.

$sel:description:UpdateEnvironmentTemplate', updateEnvironmentTemplate_description - A description of the environment template update.

$sel:name:UpdateEnvironmentTemplate', updateEnvironmentTemplate_name - The name of the environment template to update.

Request Lenses

updateEnvironmentTemplate_displayName :: Lens' UpdateEnvironmentTemplate (Maybe Text)

The name of the environment template to update as displayed in the developer interface.

updateEnvironmentTemplate_description :: Lens' UpdateEnvironmentTemplate (Maybe Text)

A description of the environment template update.

updateEnvironmentTemplate_name :: Lens' UpdateEnvironmentTemplate Text

The name of the environment template to update.

Destructuring the Response

data UpdateEnvironmentTemplateResponse

See: newUpdateEnvironmentTemplateResponse smart constructor.

Constructors

UpdateEnvironmentTemplateResponse'

Fields

httpStatus :: Int

The response's http status code.

environmentTemplate :: EnvironmentTemplate

The environment template detail data that's returned by AWS Proton.

newUpdateEnvironmentTemplateResponse

Arguments

:: Int
$sel:httpStatus:UpdateEnvironmentTemplateResponse'

-> EnvironmentTemplate
$sel:environmentTemplate:UpdateEnvironmentTemplateResponse'

-> UpdateEnvironmentTemplateResponse

Create a value of UpdateEnvironmentTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateEnvironmentTemplateResponse', updateEnvironmentTemplateResponse_httpStatus - The response's http status code.

$sel:environmentTemplate:UpdateEnvironmentTemplateResponse', updateEnvironmentTemplateResponse_environmentTemplate - The environment template detail data that's returned by AWS Proton.

Response Lenses

updateEnvironmentTemplateResponse_httpStatus :: Lens' UpdateEnvironmentTemplateResponse Int

The response's http status code.

updateEnvironmentTemplateResponse_environmentTemplate :: Lens' UpdateEnvironmentTemplateResponse EnvironmentTemplate

The environment template detail data that's returned by AWS Proton.