Network.AWS.Proton.UpdateServicePipeline

Description

Update the service pipeline.

There are four modes for updating a service pipeline as described in the following. The deploymentType field defines the mode.

[] NONE

In this mode, a deployment doesn't occur. Only the requested metadata parameters are updated.

[] CURRENT_VERSION

In this mode, the service pipeline is deployed and updated with the new spec that you provide. Only requested parameters are updated. Don’t include minor or major version parameters when you use this deployment-type.

[] MINOR_VERSION

In this mode, the service pipeline is deployed and updated with the published, recommended (latest) minor version of the current major version in use, by default. You can also specify a different minor version of the current major version in use.

[] MAJOR_VERSION

In this mode, the service pipeline is deployed and updated with the published, recommended (latest) major and minor version of the current template by default. You can also specify a different major version that is higher than the major version in use and a minor version (optional).

Creating a Request

data UpdateServicePipeline

See: newUpdateServicePipeline smart constructor.

Constructors

UpdateServicePipeline'

Fields

templateMinorVersion :: Maybe Text

The minor version of the service template that was used to create the service that the pipeline is associated with.

templateMajorVersion :: Maybe Text

The major version of the service template that was used to create the service that the pipeline is associated with.

deploymentType :: DeploymentUpdateType

The deployment type.

There are four modes for updating a service pipeline as described in the following. The deploymentType field defines the mode.

[] NONE

In this mode, a deployment doesn't occur. Only the requested metadata parameters are updated.

[] CURRENT_VERSION

In this mode, the service pipeline is deployed and updated with the new spec that you provide. Only requested parameters are updated. Don’t include minor or major version parameters when you use this deployment-type.

[] MINOR_VERSION

In this mode, the service pipeline is deployed and updated with the published, recommended (latest) minor version of the current major version in use, by default. You can also specify a different minor version of the current major version in use.

[] MAJOR_VERSION

In this mode, the service pipeline is deployed and updated with the published, recommended (latest) major and minor version of the current template, by default. You can also specify a different major version that is higher than the major version in use and a minor version (optional).

serviceName :: Text

The name of the service to that the pipeline is associated with.

spec :: Sensitive Text

The spec for the service pipeline to update.

newUpdateServicePipeline

Arguments

:: DeploymentUpdateType
$sel:deploymentType:UpdateServicePipeline'

-> Text
$sel:serviceName:UpdateServicePipeline'

-> Text
$sel:spec:UpdateServicePipeline'

-> UpdateServicePipeline

Create a value of UpdateServicePipeline with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:templateMinorVersion:UpdateServicePipeline', updateServicePipeline_templateMinorVersion - The minor version of the service template that was used to create the service that the pipeline is associated with.

$sel:templateMajorVersion:UpdateServicePipeline', updateServicePipeline_templateMajorVersion - The major version of the service template that was used to create the service that the pipeline is associated with.

$sel:deploymentType:UpdateServicePipeline', updateServicePipeline_deploymentType - The deployment type.

There are four modes for updating a service pipeline as described in the following. The deploymentType field defines the mode.

[] NONE

In this mode, a deployment doesn't occur. Only the requested metadata parameters are updated.

[] CURRENT_VERSION

In this mode, the service pipeline is deployed and updated with the new spec that you provide. Only requested parameters are updated. Don’t include minor or major version parameters when you use this deployment-type.

[] MINOR_VERSION

In this mode, the service pipeline is deployed and updated with the published, recommended (latest) minor version of the current major version in use, by default. You can also specify a different minor version of the current major version in use.

[] MAJOR_VERSION

In this mode, the service pipeline is deployed and updated with the published, recommended (latest) major and minor version of the current template, by default. You can also specify a different major version that is higher than the major version in use and a minor version (optional).

$sel:serviceName:UpdateServicePipeline', updateServicePipeline_serviceName - The name of the service to that the pipeline is associated with.

$sel:spec:UpdateServicePipeline', updateServicePipeline_spec - The spec for the service pipeline to update.

Request Lenses

updateServicePipeline_templateMinorVersion :: Lens' UpdateServicePipeline (Maybe Text)

The minor version of the service template that was used to create the service that the pipeline is associated with.

updateServicePipeline_templateMajorVersion :: Lens' UpdateServicePipeline (Maybe Text)

The major version of the service template that was used to create the service that the pipeline is associated with.

updateServicePipeline_deploymentType :: Lens' UpdateServicePipeline DeploymentUpdateType

The deployment type.

There are four modes for updating a service pipeline as described in the following. The deploymentType field defines the mode.

[] NONE

In this mode, a deployment doesn't occur. Only the requested metadata parameters are updated.

[] CURRENT_VERSION

In this mode, the service pipeline is deployed and updated with the new spec that you provide. Only requested parameters are updated. Don’t include minor or major version parameters when you use this deployment-type.

[] MINOR_VERSION

In this mode, the service pipeline is deployed and updated with the published, recommended (latest) minor version of the current major version in use, by default. You can also specify a different minor version of the current major version in use.

[] MAJOR_VERSION

In this mode, the service pipeline is deployed and updated with the published, recommended (latest) major and minor version of the current template, by default. You can also specify a different major version that is higher than the major version in use and a minor version (optional).

updateServicePipeline_serviceName :: Lens' UpdateServicePipeline Text

The name of the service to that the pipeline is associated with.

updateServicePipeline_spec :: Lens' UpdateServicePipeline Text

The spec for the service pipeline to update.

Destructuring the Response

data UpdateServicePipelineResponse

See: newUpdateServicePipelineResponse smart constructor.

Constructors

UpdateServicePipelineResponse'

Fields

httpStatus :: Int

The response's http status code.

pipeline :: ServicePipeline

The pipeline details returned by AWS Proton.

newUpdateServicePipelineResponse

Arguments

:: Int
$sel:httpStatus:UpdateServicePipelineResponse'

-> ServicePipeline
$sel:pipeline:UpdateServicePipelineResponse'

-> UpdateServicePipelineResponse

Create a value of UpdateServicePipelineResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:UpdateServicePipelineResponse', updateServicePipelineResponse_httpStatus - The response's http status code.

$sel:pipeline:UpdateServicePipelineResponse', updateServicePipelineResponse_pipeline - The pipeline details returned by AWS Proton.

Response Lenses

updateServicePipelineResponse_httpStatus :: Lens' UpdateServicePipelineResponse Int

The response's http status code.

updateServicePipelineResponse_pipeline :: Lens' UpdateServicePipelineResponse ServicePipeline

The pipeline details returned by AWS Proton.