Network.AWS.QLDB.CreateLedger

Description

Creates a new ledger in your account in the current Region.

Creating a Request

data CreateLedger

See: newCreateLedger smart constructor.

Constructors

CreateLedger'

Fields

deletionProtection :: Maybe Bool

The flag that prevents a ledger from being deleted by any user. If not provided on ledger creation, this feature is enabled (true) by default.

If deletion protection is enabled, you must first disable it before you can delete the ledger. You can disable it by calling the UpdateLedger operation to set the flag to false.

kmsKey :: Maybe Text

The key in Key Management Service (KMS) to use for encryption of data at rest in the ledger. For more information, see Encryption at rest in the Amazon QLDB Developer Guide.

Use one of the following options to specify this parameter:

AWS_OWNED_KMS_KEY: Use an KMS key that is owned and managed by Amazon Web Services on your behalf.

Undefined: By default, use an Amazon Web Services owned KMS key.

A valid symmetric customer managed KMS key: Use the specified KMS key in your account that you create, own, and manage.

Amazon QLDB does not support asymmetric keys. For more information, see Using symmetric and asymmetric keys in the Key Management Service Developer Guide.

To specify a customer managed KMS key, you can use its key ID, Amazon Resource Name (ARN), alias name, or alias ARN. When using an alias name, prefix it with "alias/". To specify a key in a different account, you must use the key ARN or alias ARN.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Alias name: alias/ExampleAlias

Alias ARN: arn:aws:kms:us-east-2:111122223333:alias/ExampleAlias

For more information, see Key identifiers (KeyId) in the Key Management Service Developer Guide.

tags :: Maybe (HashMap Text Text)

The key-value pairs to add as tags to the ledger that you want to create. Tag keys are case sensitive. Tag values are case sensitive and can be null.

name :: Text

The name of the ledger that you want to create. The name must be unique among all of the ledgers in your account in the current Region.

Naming constraints for ledger names are defined in Quotas in Amazon QLDB in the Amazon QLDB Developer Guide.

permissionsMode :: PermissionsMode

The permissions mode to assign to the ledger that you want to create. This parameter can have one of the following values:

ALLOW_ALL: A legacy permissions mode that enables access control with API-level granularity for ledgers.

This mode allows users who have the SendCommand API permission for this ledger to run all PartiQL commands (hence, ALLOW_ALL) on any tables in the specified ledger. This mode disregards any table-level or command-level IAM permissions policies that you create for the ledger.

STANDARD: (Recommended) A permissions mode that enables access control with finer granularity for ledgers, tables, and PartiQL commands.

By default, this mode denies all user requests to run any PartiQL commands on any tables in this ledger. To allow PartiQL commands to run, you must create IAM permissions policies for specific table resources and PartiQL actions, in addition to the SendCommand API permission for the ledger. For information, see Getting started with the standard permissions mode in the Amazon QLDB Developer Guide.

We strongly recommend using the STANDARD permissions mode to maximize the security of your ledger data.

newCreateLedger

Arguments

:: Text
$sel:name:CreateLedger'

-> PermissionsMode
$sel:permissionsMode:CreateLedger'

-> CreateLedger

Create a value of CreateLedger with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deletionProtection:CreateLedger', createLedger_deletionProtection - The flag that prevents a ledger from being deleted by any user. If not provided on ledger creation, this feature is enabled (true) by default.

If deletion protection is enabled, you must first disable it before you can delete the ledger. You can disable it by calling the UpdateLedger operation to set the flag to false.

$sel:kmsKey:CreateLedger', createLedger_kmsKey - The key in Key Management Service (KMS) to use for encryption of data at rest in the ledger. For more information, see Encryption at rest in the Amazon QLDB Developer Guide.

Use one of the following options to specify this parameter:

AWS_OWNED_KMS_KEY: Use an KMS key that is owned and managed by Amazon Web Services on your behalf.

Undefined: By default, use an Amazon Web Services owned KMS key.

A valid symmetric customer managed KMS key: Use the specified KMS key in your account that you create, own, and manage.

Amazon QLDB does not support asymmetric keys. For more information, see Using symmetric and asymmetric keys in the Key Management Service Developer Guide.

To specify a customer managed KMS key, you can use its key ID, Amazon Resource Name (ARN), alias name, or alias ARN. When using an alias name, prefix it with "alias/". To specify a key in a different account, you must use the key ARN or alias ARN.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Alias name: alias/ExampleAlias

Alias ARN: arn:aws:kms:us-east-2:111122223333:alias/ExampleAlias

For more information, see Key identifiers (KeyId) in the Key Management Service Developer Guide.

$sel:tags:CreateLedger', createLedger_tags - The key-value pairs to add as tags to the ledger that you want to create. Tag keys are case sensitive. Tag values are case sensitive and can be null.

$sel:name:CreateLedger', createLedger_name - The name of the ledger that you want to create. The name must be unique among all of the ledgers in your account in the current Region.

Naming constraints for ledger names are defined in Quotas in Amazon QLDB in the Amazon QLDB Developer Guide.

$sel:permissionsMode:CreateLedger', createLedger_permissionsMode - The permissions mode to assign to the ledger that you want to create. This parameter can have one of the following values:

ALLOW_ALL: A legacy permissions mode that enables access control with API-level granularity for ledgers.

This mode allows users who have the SendCommand API permission for this ledger to run all PartiQL commands (hence, ALLOW_ALL) on any tables in the specified ledger. This mode disregards any table-level or command-level IAM permissions policies that you create for the ledger.

STANDARD: (Recommended) A permissions mode that enables access control with finer granularity for ledgers, tables, and PartiQL commands.

By default, this mode denies all user requests to run any PartiQL commands on any tables in this ledger. To allow PartiQL commands to run, you must create IAM permissions policies for specific table resources and PartiQL actions, in addition to the SendCommand API permission for the ledger. For information, see Getting started with the standard permissions mode in the Amazon QLDB Developer Guide.

We strongly recommend using the STANDARD permissions mode to maximize the security of your ledger data.

Request Lenses

createLedger_deletionProtection :: Lens' CreateLedger (Maybe Bool)

The flag that prevents a ledger from being deleted by any user. If not provided on ledger creation, this feature is enabled (true) by default.

If deletion protection is enabled, you must first disable it before you can delete the ledger. You can disable it by calling the UpdateLedger operation to set the flag to false.

createLedger_kmsKey :: Lens' CreateLedger (Maybe Text)

The key in Key Management Service (KMS) to use for encryption of data at rest in the ledger. For more information, see Encryption at rest in the Amazon QLDB Developer Guide.

Use one of the following options to specify this parameter:

AWS_OWNED_KMS_KEY: Use an KMS key that is owned and managed by Amazon Web Services on your behalf.

Undefined: By default, use an Amazon Web Services owned KMS key.

A valid symmetric customer managed KMS key: Use the specified KMS key in your account that you create, own, and manage.

Amazon QLDB does not support asymmetric keys. For more information, see Using symmetric and asymmetric keys in the Key Management Service Developer Guide.

To specify a customer managed KMS key, you can use its key ID, Amazon Resource Name (ARN), alias name, or alias ARN. When using an alias name, prefix it with "alias/". To specify a key in a different account, you must use the key ARN or alias ARN.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Alias name: alias/ExampleAlias

Alias ARN: arn:aws:kms:us-east-2:111122223333:alias/ExampleAlias

For more information, see Key identifiers (KeyId) in the Key Management Service Developer Guide.

createLedger_tags :: Lens' CreateLedger (Maybe (HashMap Text Text))

The key-value pairs to add as tags to the ledger that you want to create. Tag keys are case sensitive. Tag values are case sensitive and can be null.

createLedger_name :: Lens' CreateLedger Text

The name of the ledger that you want to create. The name must be unique among all of the ledgers in your account in the current Region.

Naming constraints for ledger names are defined in Quotas in Amazon QLDB in the Amazon QLDB Developer Guide.

createLedger_permissionsMode :: Lens' CreateLedger PermissionsMode

The permissions mode to assign to the ledger that you want to create. This parameter can have one of the following values:

ALLOW_ALL: A legacy permissions mode that enables access control with API-level granularity for ledgers.

This mode allows users who have the SendCommand API permission for this ledger to run all PartiQL commands (hence, ALLOW_ALL) on any tables in the specified ledger. This mode disregards any table-level or command-level IAM permissions policies that you create for the ledger.

STANDARD: (Recommended) A permissions mode that enables access control with finer granularity for ledgers, tables, and PartiQL commands.

By default, this mode denies all user requests to run any PartiQL commands on any tables in this ledger. To allow PartiQL commands to run, you must create IAM permissions policies for specific table resources and PartiQL actions, in addition to the SendCommand API permission for the ledger. For information, see Getting started with the standard permissions mode in the Amazon QLDB Developer Guide.

We strongly recommend using the STANDARD permissions mode to maximize the security of your ledger data.

Destructuring the Response

data CreateLedgerResponse

See: newCreateLedgerResponse smart constructor.

Constructors

CreateLedgerResponse'

Fields

deletionProtection :: Maybe Bool

The flag that prevents a ledger from being deleted by any user. If not provided on ledger creation, this feature is enabled (true) by default.

If deletion protection is enabled, you must first disable it before you can delete the ledger. You can disable it by calling the UpdateLedger operation to set the flag to false.

permissionsMode :: Maybe PermissionsMode

The permissions mode of the ledger that you created.

arn :: Maybe Text

The Amazon Resource Name (ARN) for the ledger.

name :: Maybe Text

The name of the ledger.

kmsKeyArn :: Maybe Text

The ARN of the customer managed KMS key that the ledger uses for encryption at rest. If this parameter is undefined, the ledger uses an Amazon Web Services owned KMS key for encryption.

state :: Maybe LedgerState

The current status of the ledger.

creationDateTime :: Maybe POSIX

The date and time, in epoch time format, when the ledger was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

httpStatus :: Int

The response's http status code.

newCreateLedgerResponse

Arguments

:: Int
$sel:httpStatus:CreateLedgerResponse'

-> CreateLedgerResponse

Create a value of CreateLedgerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deletionProtection:CreateLedger', createLedgerResponse_deletionProtection - The flag that prevents a ledger from being deleted by any user. If not provided on ledger creation, this feature is enabled (true) by default.

If deletion protection is enabled, you must first disable it before you can delete the ledger. You can disable it by calling the UpdateLedger operation to set the flag to false.

$sel:permissionsMode:CreateLedger', createLedgerResponse_permissionsMode - The permissions mode of the ledger that you created.

$sel:arn:CreateLedgerResponse', createLedgerResponse_arn - The Amazon Resource Name (ARN) for the ledger.

$sel:name:CreateLedger', createLedgerResponse_name - The name of the ledger.

$sel:kmsKeyArn:CreateLedgerResponse', createLedgerResponse_kmsKeyArn - The ARN of the customer managed KMS key that the ledger uses for encryption at rest. If this parameter is undefined, the ledger uses an Amazon Web Services owned KMS key for encryption.

$sel:state:CreateLedgerResponse', createLedgerResponse_state - The current status of the ledger.

$sel:creationDateTime:CreateLedgerResponse', createLedgerResponse_creationDateTime - The date and time, in epoch time format, when the ledger was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

$sel:httpStatus:CreateLedgerResponse', createLedgerResponse_httpStatus - The response's http status code.

Response Lenses

createLedgerResponse_deletionProtection :: Lens' CreateLedgerResponse (Maybe Bool)

The flag that prevents a ledger from being deleted by any user. If not provided on ledger creation, this feature is enabled (true) by default.

If deletion protection is enabled, you must first disable it before you can delete the ledger. You can disable it by calling the UpdateLedger operation to set the flag to false.

createLedgerResponse_permissionsMode :: Lens' CreateLedgerResponse (Maybe PermissionsMode)

The permissions mode of the ledger that you created.

createLedgerResponse_arn :: Lens' CreateLedgerResponse (Maybe Text)

The Amazon Resource Name (ARN) for the ledger.

createLedgerResponse_name :: Lens' CreateLedgerResponse (Maybe Text)

The name of the ledger.

createLedgerResponse_kmsKeyArn :: Lens' CreateLedgerResponse (Maybe Text)

The ARN of the customer managed KMS key that the ledger uses for encryption at rest. If this parameter is undefined, the ledger uses an Amazon Web Services owned KMS key for encryption.

createLedgerResponse_state :: Lens' CreateLedgerResponse (Maybe LedgerState)

The current status of the ledger.

createLedgerResponse_creationDateTime :: Lens' CreateLedgerResponse (Maybe UTCTime)

The date and time, in epoch time format, when the ledger was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

createLedgerResponse_httpStatus :: Lens' CreateLedgerResponse Int

The response's http status code.