Network.AWS.QLDB.DescribeJournalKinesisStream

Description

Returns detailed information about a given Amazon QLDB journal stream. The output includes the Amazon Resource Name (ARN), stream name, current status, creation time, and the parameters of the original stream creation request.

This action does not return any expired journal streams. For more information, see Expiration for terminal streams in the Amazon QLDB Developer Guide.

Creating a Request

data DescribeJournalKinesisStream

See: newDescribeJournalKinesisStream smart constructor.

Constructors

DescribeJournalKinesisStream'

Fields

ledgerName :: Text

The name of the ledger.

streamId :: Text

The UUID (represented in Base62-encoded text) of the QLDB journal stream to describe.

newDescribeJournalKinesisStream

Arguments

:: Text
$sel:ledgerName:DescribeJournalKinesisStream'

-> Text
$sel:streamId:DescribeJournalKinesisStream'

-> DescribeJournalKinesisStream

Create a value of DescribeJournalKinesisStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ledgerName:DescribeJournalKinesisStream', describeJournalKinesisStream_ledgerName - The name of the ledger.

$sel:streamId:DescribeJournalKinesisStream', describeJournalKinesisStream_streamId - The UUID (represented in Base62-encoded text) of the QLDB journal stream to describe.

Request Lenses

describeJournalKinesisStream_ledgerName :: Lens' DescribeJournalKinesisStream Text

The name of the ledger.

describeJournalKinesisStream_streamId :: Lens' DescribeJournalKinesisStream Text

The UUID (represented in Base62-encoded text) of the QLDB journal stream to describe.

Destructuring the Response

data DescribeJournalKinesisStreamResponse

See: newDescribeJournalKinesisStreamResponse smart constructor.

Constructors

DescribeJournalKinesisStreamResponse'

Fields

stream :: Maybe JournalKinesisStreamDescription

Information about the QLDB journal stream returned by a DescribeJournalS3Export request.

httpStatus :: Int

The response's http status code.

newDescribeJournalKinesisStreamResponse

Arguments

:: Int
$sel:httpStatus:DescribeJournalKinesisStreamResponse'

-> DescribeJournalKinesisStreamResponse

Create a value of DescribeJournalKinesisStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:stream:DescribeJournalKinesisStreamResponse', describeJournalKinesisStreamResponse_stream - Information about the QLDB journal stream returned by a DescribeJournalS3Export request.

$sel:httpStatus:DescribeJournalKinesisStreamResponse', describeJournalKinesisStreamResponse_httpStatus - The response's http status code.

Response Lenses

describeJournalKinesisStreamResponse_stream :: Lens' DescribeJournalKinesisStreamResponse (Maybe JournalKinesisStreamDescription)

Information about the QLDB journal stream returned by a DescribeJournalS3Export request.

describeJournalKinesisStreamResponse_httpStatus :: Lens' DescribeJournalKinesisStreamResponse Int

The response's http status code.