Network.AWS.QLDB.DescribeJournalS3Export

Description

Returns information about a journal export job, including the ledger name, export ID, creation time, current status, and the parameters of the original export creation request.

This action does not return any expired export jobs. For more information, see Export job expiration in the Amazon QLDB Developer Guide.

If the export job with the given ExportId doesn't exist, then throws ResourceNotFoundException.

If the ledger with the given Name doesn't exist, then throws ResourceNotFoundException.

Creating a Request

data DescribeJournalS3Export

See: newDescribeJournalS3Export smart constructor.

Constructors

DescribeJournalS3Export'

Fields

name :: Text

The name of the ledger.

exportId :: Text

The UUID (represented in Base62-encoded text) of the journal export job to describe.

newDescribeJournalS3Export

Arguments

:: Text
$sel:name:DescribeJournalS3Export'

-> Text
$sel:exportId:DescribeJournalS3Export'

-> DescribeJournalS3Export

Create a value of DescribeJournalS3Export with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:DescribeJournalS3Export', describeJournalS3Export_name - The name of the ledger.

$sel:exportId:DescribeJournalS3Export', describeJournalS3Export_exportId - The UUID (represented in Base62-encoded text) of the journal export job to describe.

Request Lenses

describeJournalS3Export_name :: Lens' DescribeJournalS3Export Text

The name of the ledger.

describeJournalS3Export_exportId :: Lens' DescribeJournalS3Export Text

The UUID (represented in Base62-encoded text) of the journal export job to describe.

Destructuring the Response

data DescribeJournalS3ExportResponse

See: newDescribeJournalS3ExportResponse smart constructor.

Constructors

DescribeJournalS3ExportResponse'

Fields

httpStatus :: Int

The response's http status code.

exportDescription :: JournalS3ExportDescription

Information about the journal export job returned by a DescribeJournalS3Export request.

newDescribeJournalS3ExportResponse

Arguments

:: Int
$sel:httpStatus:DescribeJournalS3ExportResponse'

-> JournalS3ExportDescription
$sel:exportDescription:DescribeJournalS3ExportResponse'

-> DescribeJournalS3ExportResponse

Create a value of DescribeJournalS3ExportResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeJournalS3ExportResponse', describeJournalS3ExportResponse_httpStatus - The response's http status code.

$sel:exportDescription:DescribeJournalS3ExportResponse', describeJournalS3ExportResponse_exportDescription - Information about the journal export job returned by a DescribeJournalS3Export request.

Response Lenses

describeJournalS3ExportResponse_httpStatus :: Lens' DescribeJournalS3ExportResponse Int

The response's http status code.

describeJournalS3ExportResponse_exportDescription :: Lens' DescribeJournalS3ExportResponse JournalS3ExportDescription

Information about the journal export job returned by a DescribeJournalS3Export request.