Network.AWS.QLDB.ExportJournalToS

Description

Exports journal contents within a date and time range from a ledger into a specified Amazon Simple Storage Service (Amazon S3) bucket. The data is written as files in Amazon Ion format.

If the ledger with the given Name doesn't exist, then throws ResourceNotFoundException.

If the ledger with the given Name is in CREATING status, then throws ResourcePreconditionNotMetException.

You can initiate up to two concurrent journal export requests for each ledger. Beyond this limit, journal export requests throw LimitExceededException.

Creating a Request

data ExportJournalToS

See: newExportJournalToS smart constructor.

Constructors

ExportJournalToS'

Fields

name :: Text

The name of the ledger.

inclusiveStartTime :: POSIX

The inclusive start date and time for the range of journal contents to export.

The InclusiveStartTime must be in ISO 8601 date and time format and in Universal Coordinated Time (UTC). For example: 2019-06-13T21:36:34Z.

The InclusiveStartTime must be before ExclusiveEndTime.

If you provide an InclusiveStartTime that is before the ledger's CreationDateTime, Amazon QLDB defaults it to the ledger's CreationDateTime.

exclusiveEndTime :: POSIX

The exclusive end date and time for the range of journal contents to export.

The ExclusiveEndTime must be in ISO 8601 date and time format and in Universal Coordinated Time (UTC). For example: 2019-06-13T21:36:34Z.

The ExclusiveEndTime must be less than or equal to the current UTC date and time.

s3ExportConfiguration :: S3ExportConfiguration

The configuration settings of the Amazon S3 bucket destination for your export request.

roleArn :: Text

The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal export job to do the following:

Write objects into your Amazon Simple Storage Service (Amazon S3) bucket.

(Optional) Use your customer master key (CMK) in Key Management Service (KMS) for server-side encryption of your exported data.

newExportJournalToS

Arguments

:: Text
$sel:name:ExportJournalToS'

-> UTCTime
$sel:inclusiveStartTime:ExportJournalToS'

-> UTCTime
$sel:exclusiveEndTime:ExportJournalToS'

-> S3ExportConfiguration
$sel:s3ExportConfiguration:ExportJournalToS'

-> Text
$sel:roleArn:ExportJournalToS'

-> ExportJournalToS

Create a value of ExportJournalToS with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ExportJournalToS', exportJournalToS_name - The name of the ledger.

$sel:inclusiveStartTime:ExportJournalToS', exportJournalToS_inclusiveStartTime - The inclusive start date and time for the range of journal contents to export.

The InclusiveStartTime must be in ISO 8601 date and time format and in Universal Coordinated Time (UTC). For example: 2019-06-13T21:36:34Z.

The InclusiveStartTime must be before ExclusiveEndTime.

If you provide an InclusiveStartTime that is before the ledger's CreationDateTime, Amazon QLDB defaults it to the ledger's CreationDateTime.

$sel:exclusiveEndTime:ExportJournalToS', exportJournalToS_exclusiveEndTime - The exclusive end date and time for the range of journal contents to export.

The ExclusiveEndTime must be in ISO 8601 date and time format and in Universal Coordinated Time (UTC). For example: 2019-06-13T21:36:34Z.

The ExclusiveEndTime must be less than or equal to the current UTC date and time.

$sel:s3ExportConfiguration:ExportJournalToS', exportJournalToS_s3ExportConfiguration - The configuration settings of the Amazon S3 bucket destination for your export request.

$sel:roleArn:ExportJournalToS', exportJournalToS_roleArn - The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal export job to do the following:

Write objects into your Amazon Simple Storage Service (Amazon S3) bucket.

(Optional) Use your customer master key (CMK) in Key Management Service (KMS) for server-side encryption of your exported data.

Request Lenses

exportJournalToS_name :: Lens' ExportJournalToS Text

The name of the ledger.

exportJournalToS_inclusiveStartTime :: Lens' ExportJournalToS UTCTime

The inclusive start date and time for the range of journal contents to export.

The InclusiveStartTime must be in ISO 8601 date and time format and in Universal Coordinated Time (UTC). For example: 2019-06-13T21:36:34Z.

The InclusiveStartTime must be before ExclusiveEndTime.

If you provide an InclusiveStartTime that is before the ledger's CreationDateTime, Amazon QLDB defaults it to the ledger's CreationDateTime.

exportJournalToS_exclusiveEndTime :: Lens' ExportJournalToS UTCTime

The exclusive end date and time for the range of journal contents to export.

The ExclusiveEndTime must be in ISO 8601 date and time format and in Universal Coordinated Time (UTC). For example: 2019-06-13T21:36:34Z.

The ExclusiveEndTime must be less than or equal to the current UTC date and time.

exportJournalToS_s3ExportConfiguration :: Lens' ExportJournalToS S3ExportConfiguration

The configuration settings of the Amazon S3 bucket destination for your export request.

exportJournalToS_roleArn :: Lens' ExportJournalToS Text

The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal export job to do the following:

Write objects into your Amazon Simple Storage Service (Amazon S3) bucket.

(Optional) Use your customer master key (CMK) in Key Management Service (KMS) for server-side encryption of your exported data.

Destructuring the Response

data ExportJournalToSResponse

See: newExportJournalToSResponse smart constructor.

Constructors

ExportJournalToSResponse'

Fields

httpStatus :: Int

The response's http status code.

exportId :: Text

The UUID (represented in Base62-encoded text) that QLDB assigns to each journal export job.

To describe your export request and check the status of the job, you can use ExportId to call DescribeJournalS3Export.

newExportJournalToSResponse

Arguments

:: Int
$sel:httpStatus:ExportJournalToSResponse'

-> Text
$sel:exportId:ExportJournalToSResponse'

-> ExportJournalToSResponse

Create a value of ExportJournalToSResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:ExportJournalToSResponse', exportJournalToSResponse_httpStatus - The response's http status code.

$sel:exportId:ExportJournalToSResponse', exportJournalToSResponse_exportId - The UUID (represented in Base62-encoded text) that QLDB assigns to each journal export job.

To describe your export request and check the status of the job, you can use ExportId to call DescribeJournalS3Export.

Response Lenses

exportJournalToSResponse_httpStatus :: Lens' ExportJournalToSResponse Int

The response's http status code.

exportJournalToSResponse_exportId :: Lens' ExportJournalToSResponse Text

The UUID (represented in Base62-encoded text) that QLDB assigns to each journal export job.

To describe your export request and check the status of the job, you can use ExportId to call DescribeJournalS3Export.