Network.AWS.QLDB.GetRevision

Description

Returns a revision data object for a specified document ID and block address. Also returns a proof of the specified revision for verification if DigestTipAddress is provided.

Creating a Request

data GetRevision

See: newGetRevision smart constructor.

Constructors

GetRevision'

Fields

digestTipAddress :: Maybe (Sensitive ValueHolder)

The latest block location covered by the digest for which to request a proof. An address is an Amazon Ion structure that has two fields: strandId and sequenceNo.

For example: {strandId:"BlFTjlSXze9BIh1KOszcE3",sequenceNo:49}.

name :: Text

The name of the ledger.

blockAddress :: Sensitive ValueHolder

The block location of the document revision to be verified. An address is an Amazon Ion structure that has two fields: strandId and sequenceNo.

For example: {strandId:"BlFTjlSXze9BIh1KOszcE3",sequenceNo:14}.

documentId :: Text

The UUID (represented in Base62-encoded text) of the document to be verified.

newGetRevision

Arguments

:: Text
$sel:name:GetRevision'

-> ValueHolder
$sel:blockAddress:GetRevision'

-> Text
$sel:documentId:GetRevision'

-> GetRevision

Create a value of GetRevision with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:digestTipAddress:GetRevision', getRevision_digestTipAddress - The latest block location covered by the digest for which to request a proof. An address is an Amazon Ion structure that has two fields: strandId and sequenceNo.

For example: {strandId:"BlFTjlSXze9BIh1KOszcE3",sequenceNo:49}.

$sel:name:GetRevision', getRevision_name - The name of the ledger.

$sel:blockAddress:GetRevision', getRevision_blockAddress - The block location of the document revision to be verified. An address is an Amazon Ion structure that has two fields: strandId and sequenceNo.

For example: {strandId:"BlFTjlSXze9BIh1KOszcE3",sequenceNo:14}.

$sel:documentId:GetRevision', getRevision_documentId - The UUID (represented in Base62-encoded text) of the document to be verified.

Request Lenses

getRevision_digestTipAddress :: Lens' GetRevision (Maybe ValueHolder)

The latest block location covered by the digest for which to request a proof. An address is an Amazon Ion structure that has two fields: strandId and sequenceNo.

For example: {strandId:"BlFTjlSXze9BIh1KOszcE3",sequenceNo:49}.

getRevision_name :: Lens' GetRevision Text

The name of the ledger.

getRevision_blockAddress :: Lens' GetRevision ValueHolder

The block location of the document revision to be verified. An address is an Amazon Ion structure that has two fields: strandId and sequenceNo.

For example: {strandId:"BlFTjlSXze9BIh1KOszcE3",sequenceNo:14}.

getRevision_documentId :: Lens' GetRevision Text

The UUID (represented in Base62-encoded text) of the document to be verified.

Destructuring the Response

data GetRevisionResponse

See: newGetRevisionResponse smart constructor.

Constructors

GetRevisionResponse'

Fields

proof :: Maybe (Sensitive ValueHolder)

The proof object in Amazon Ion format returned by a GetRevision request. A proof contains the list of hash values that are required to recalculate the specified digest using a Merkle tree, starting with the specified document revision.

httpStatus :: Int

The response's http status code.

revision :: Sensitive ValueHolder

The document revision data object in Amazon Ion format.

newGetRevisionResponse

Arguments

:: Int
$sel:httpStatus:GetRevisionResponse'

-> ValueHolder
$sel:revision:GetRevisionResponse'

-> GetRevisionResponse

Create a value of GetRevisionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:proof:GetRevisionResponse', getRevisionResponse_proof - The proof object in Amazon Ion format returned by a GetRevision request. A proof contains the list of hash values that are required to recalculate the specified digest using a Merkle tree, starting with the specified document revision.

$sel:httpStatus:GetRevisionResponse', getRevisionResponse_httpStatus - The response's http status code.

$sel:revision:GetRevisionResponse', getRevisionResponse_revision - The document revision data object in Amazon Ion format.

Response Lenses

getRevisionResponse_proof :: Lens' GetRevisionResponse (Maybe ValueHolder)

The proof object in Amazon Ion format returned by a GetRevision request. A proof contains the list of hash values that are required to recalculate the specified digest using a Merkle tree, starting with the specified document revision.

getRevisionResponse_httpStatus :: Lens' GetRevisionResponse Int

The response's http status code.

getRevisionResponse_revision :: Lens' GetRevisionResponse ValueHolder

The document revision data object in Amazon Ion format.