Network.AWS.QLDB.Lens

Description

Operations

UpdateLedger

updateLedger_deletionProtection :: Lens' UpdateLedger (Maybe Bool)

The flag that prevents a ledger from being deleted by any user. If not provided on ledger creation, this feature is enabled (true) by default.

If deletion protection is enabled, you must first disable it before you can delete the ledger. You can disable it by calling the UpdateLedger operation to set the flag to false.

updateLedger_kmsKey :: Lens' UpdateLedger (Maybe Text)

The key in Key Management Service (KMS) to use for encryption of data at rest in the ledger. For more information, see Encryption at rest in the Amazon QLDB Developer Guide.

Use one of the following options to specify this parameter:

AWS_OWNED_KMS_KEY: Use an KMS key that is owned and managed by Amazon Web Services on your behalf.

Undefined: Make no changes to the KMS key of the ledger.

A valid symmetric customer managed KMS key: Use the specified KMS key in your account that you create, own, and manage.

Amazon QLDB does not support asymmetric keys. For more information, see Using symmetric and asymmetric keys in the Key Management Service Developer Guide.

To specify a customer managed KMS key, you can use its key ID, Amazon Resource Name (ARN), alias name, or alias ARN. When using an alias name, prefix it with "alias/". To specify a key in a different account, you must use the key ARN or alias ARN.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Alias name: alias/ExampleAlias

Alias ARN: arn:aws:kms:us-east-2:111122223333:alias/ExampleAlias

For more information, see Key identifiers (KeyId) in the Key Management Service Developer Guide.

updateLedger_name :: Lens' UpdateLedger Text

The name of the ledger.

updateLedgerResponse_state :: Lens' UpdateLedgerResponse (Maybe LedgerState)

The current status of the ledger.

updateLedgerResponse_deletionProtection :: Lens' UpdateLedgerResponse (Maybe Bool)

The flag that prevents a ledger from being deleted by any user. If not provided on ledger creation, this feature is enabled (true) by default.

If deletion protection is enabled, you must first disable it before you can delete the ledger. You can disable it by calling the UpdateLedger operation to set the flag to false.

updateLedgerResponse_arn :: Lens' UpdateLedgerResponse (Maybe Text)

The Amazon Resource Name (ARN) for the ledger.

updateLedgerResponse_encryptionDescription :: Lens' UpdateLedgerResponse (Maybe LedgerEncryptionDescription)

Information about the encryption of data at rest in the ledger. This includes the current status, the KMS key, and when the key became inaccessible (in the case of an error).

updateLedgerResponse_name :: Lens' UpdateLedgerResponse (Maybe Text)

The name of the ledger.

updateLedgerResponse_creationDateTime :: Lens' UpdateLedgerResponse (Maybe UTCTime)

The date and time, in epoch time format, when the ledger was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

updateLedgerResponse_httpStatus :: Lens' UpdateLedgerResponse Int

The response's http status code.

DeleteLedger

deleteLedger_name :: Lens' DeleteLedger Text

The name of the ledger that you want to delete.

ListJournalKinesisStreamsForLedger

listJournalKinesisStreamsForLedger_nextToken :: Lens' ListJournalKinesisStreamsForLedger (Maybe Text)

A pagination token, indicating that you want to retrieve the next page of results. If you received a value for NextToken in the response from a previous ListJournalKinesisStreamsForLedger call, you should use that value as input here.

listJournalKinesisStreamsForLedger_maxResults :: Lens' ListJournalKinesisStreamsForLedger (Maybe Natural)

The maximum number of results to return in a single ListJournalKinesisStreamsForLedger request. (The actual number of results returned might be fewer.)

listJournalKinesisStreamsForLedger_ledgerName :: Lens' ListJournalKinesisStreamsForLedger Text

The name of the ledger.

listJournalKinesisStreamsForLedgerResponse_nextToken :: Lens' ListJournalKinesisStreamsForLedgerResponse (Maybe Text)

If NextToken is empty, the last page of results has been processed and there are no more results to be retrieved.

If NextToken is not empty, more results are available. To retrieve the next page of results, use the value of NextToken in a subsequent ListJournalKinesisStreamsForLedger call.

listJournalKinesisStreamsForLedgerResponse_streams :: Lens' ListJournalKinesisStreamsForLedgerResponse (Maybe [JournalKinesisStreamDescription])

The array of QLDB journal stream descriptors that are associated with the given ledger.

listJournalKinesisStreamsForLedgerResponse_httpStatus :: Lens' ListJournalKinesisStreamsForLedgerResponse Int

The response's http status code.

ListTagsForResource

listTagsForResource_resourceArn :: Lens' ListTagsForResource Text

The Amazon Resource Name (ARN) for which to list the tags. For example:

arn:aws:qldb:us-east-1:123456789012:ledger/exampleLedger

listTagsForResourceResponse_tags :: Lens' ListTagsForResourceResponse (Maybe (HashMap Text Text))

The tags that are currently associated with the specified Amazon QLDB resource.

listTagsForResourceResponse_httpStatus :: Lens' ListTagsForResourceResponse Int

The response's http status code.

GetRevision

getRevision_digestTipAddress :: Lens' GetRevision (Maybe ValueHolder)

The latest block location covered by the digest for which to request a proof. An address is an Amazon Ion structure that has two fields: strandId and sequenceNo.

For example: {strandId:"BlFTjlSXze9BIh1KOszcE3",sequenceNo:49}.

getRevision_name :: Lens' GetRevision Text

The name of the ledger.

getRevision_blockAddress :: Lens' GetRevision ValueHolder

The block location of the document revision to be verified. An address is an Amazon Ion structure that has two fields: strandId and sequenceNo.

For example: {strandId:"BlFTjlSXze9BIh1KOszcE3",sequenceNo:14}.

getRevision_documentId :: Lens' GetRevision Text

The UUID (represented in Base62-encoded text) of the document to be verified.

getRevisionResponse_proof :: Lens' GetRevisionResponse (Maybe ValueHolder)

The proof object in Amazon Ion format returned by a GetRevision request. A proof contains the list of hash values that are required to recalculate the specified digest using a Merkle tree, starting with the specified document revision.

getRevisionResponse_httpStatus :: Lens' GetRevisionResponse Int

The response's http status code.

getRevisionResponse_revision :: Lens' GetRevisionResponse ValueHolder

The document revision data object in Amazon Ion format.

DescribeLedger

describeLedger_name :: Lens' DescribeLedger Text

The name of the ledger that you want to describe.

describeLedgerResponse_state :: Lens' DescribeLedgerResponse (Maybe LedgerState)

The current status of the ledger.

describeLedgerResponse_deletionProtection :: Lens' DescribeLedgerResponse (Maybe Bool)

The flag that prevents a ledger from being deleted by any user. If not provided on ledger creation, this feature is enabled (true) by default.

If deletion protection is enabled, you must first disable it before you can delete the ledger. You can disable it by calling the UpdateLedger operation to set the flag to false.

describeLedgerResponse_arn :: Lens' DescribeLedgerResponse (Maybe Text)

The Amazon Resource Name (ARN) for the ledger.

describeLedgerResponse_encryptionDescription :: Lens' DescribeLedgerResponse (Maybe LedgerEncryptionDescription)

Information about the encryption of data at rest in the ledger. This includes the current status, the KMS key, and when the key became inaccessible (in the case of an error).

describeLedgerResponse_name :: Lens' DescribeLedgerResponse (Maybe Text)

The name of the ledger.

describeLedgerResponse_creationDateTime :: Lens' DescribeLedgerResponse (Maybe UTCTime)

The date and time, in epoch time format, when the ledger was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

describeLedgerResponse_permissionsMode :: Lens' DescribeLedgerResponse (Maybe PermissionsMode)

The permissions mode of the ledger.

describeLedgerResponse_httpStatus :: Lens' DescribeLedgerResponse Int

The response's http status code.

CancelJournalKinesisStream

cancelJournalKinesisStream_ledgerName :: Lens' CancelJournalKinesisStream Text

The name of the ledger.

cancelJournalKinesisStream_streamId :: Lens' CancelJournalKinesisStream Text

The UUID (represented in Base62-encoded text) of the QLDB journal stream to be canceled.

cancelJournalKinesisStreamResponse_streamId :: Lens' CancelJournalKinesisStreamResponse (Maybe Text)

The UUID (Base62-encoded text) of the canceled QLDB journal stream.

cancelJournalKinesisStreamResponse_httpStatus :: Lens' CancelJournalKinesisStreamResponse Int

The response's http status code.

ExportJournalToS3

exportJournalToS3_name :: Lens' ExportJournalToS3 Text

The name of the ledger.

exportJournalToS3_inclusiveStartTime :: Lens' ExportJournalToS3 UTCTime

The inclusive start date and time for the range of journal contents to export.

The InclusiveStartTime must be in ISO 8601 date and time format and in Universal Coordinated Time (UTC). For example: 2019-06-13T21:36:34Z.

The InclusiveStartTime must be before ExclusiveEndTime.

If you provide an InclusiveStartTime that is before the ledger's CreationDateTime, Amazon QLDB defaults it to the ledger's CreationDateTime.

exportJournalToS3_exclusiveEndTime :: Lens' ExportJournalToS3 UTCTime

The exclusive end date and time for the range of journal contents to export.

The ExclusiveEndTime must be in ISO 8601 date and time format and in Universal Coordinated Time (UTC). For example: 2019-06-13T21:36:34Z.

The ExclusiveEndTime must be less than or equal to the current UTC date and time.

exportJournalToS3_s3ExportConfiguration :: Lens' ExportJournalToS3 S3ExportConfiguration

The configuration settings of the Amazon S3 bucket destination for your export request.

exportJournalToS3_roleArn :: Lens' ExportJournalToS3 Text

The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal export job to do the following:

Write objects into your Amazon Simple Storage Service (Amazon S3) bucket.

(Optional) Use your customer master key (CMK) in Key Management Service (KMS) for server-side encryption of your exported data.

exportJournalToS3Response_httpStatus :: Lens' ExportJournalToS3Response Int

The response's http status code.

exportJournalToS3Response_exportId :: Lens' ExportJournalToS3Response Text

The UUID (represented in Base62-encoded text) that QLDB assigns to each journal export job.

To describe your export request and check the status of the job, you can use ExportId to call DescribeJournalS3Export.

StreamJournalToKinesis

streamJournalToKinesis_exclusiveEndTime :: Lens' StreamJournalToKinesis (Maybe UTCTime)

The exclusive date and time that specifies when the stream ends. If you don't define this parameter, the stream runs indefinitely until you cancel it.

The ExclusiveEndTime must be in ISO 8601 date and time format and in Universal Coordinated Time (UTC). For example: 2019-06-13T21:36:34Z.

streamJournalToKinesis_tags :: Lens' StreamJournalToKinesis (Maybe (HashMap Text Text))

The key-value pairs to add as tags to the stream that you want to create. Tag keys are case sensitive. Tag values are case sensitive and can be null.

streamJournalToKinesis_ledgerName :: Lens' StreamJournalToKinesis Text

The name of the ledger.

streamJournalToKinesis_roleArn :: Lens' StreamJournalToKinesis Text

The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal stream to write data records to a Kinesis Data Streams resource.

streamJournalToKinesis_inclusiveStartTime :: Lens' StreamJournalToKinesis UTCTime

The inclusive start date and time from which to start streaming journal data. This parameter must be in ISO 8601 date and time format and in Universal Coordinated Time (UTC). For example: 2019-06-13T21:36:34Z.

The InclusiveStartTime cannot be in the future and must be before ExclusiveEndTime.

If you provide an InclusiveStartTime that is before the ledger's CreationDateTime, QLDB effectively defaults it to the ledger's CreationDateTime.

streamJournalToKinesis_kinesisConfiguration :: Lens' StreamJournalToKinesis KinesisConfiguration

The configuration settings of the Kinesis Data Streams destination for your stream request.

streamJournalToKinesis_streamName :: Lens' StreamJournalToKinesis Text

The name that you want to assign to the QLDB journal stream. User-defined names can help identify and indicate the purpose of a stream.

Your stream name must be unique among other active streams for a given ledger. Stream names have the same naming constraints as ledger names, as defined in Quotas in Amazon QLDB in the Amazon QLDB Developer Guide.

streamJournalToKinesisResponse_streamId :: Lens' StreamJournalToKinesisResponse (Maybe Text)

The UUID (represented in Base62-encoded text) that QLDB assigns to each QLDB journal stream.

streamJournalToKinesisResponse_httpStatus :: Lens' StreamJournalToKinesisResponse Int

The response's http status code.

CreateLedger

createLedger_deletionProtection :: Lens' CreateLedger (Maybe Bool)

The flag that prevents a ledger from being deleted by any user. If not provided on ledger creation, this feature is enabled (true) by default.

If deletion protection is enabled, you must first disable it before you can delete the ledger. You can disable it by calling the UpdateLedger operation to set the flag to false.

createLedger_kmsKey :: Lens' CreateLedger (Maybe Text)

The key in Key Management Service (KMS) to use for encryption of data at rest in the ledger. For more information, see Encryption at rest in the Amazon QLDB Developer Guide.

Use one of the following options to specify this parameter:

AWS_OWNED_KMS_KEY: Use an KMS key that is owned and managed by Amazon Web Services on your behalf.

Undefined: By default, use an Amazon Web Services owned KMS key.

A valid symmetric customer managed KMS key: Use the specified KMS key in your account that you create, own, and manage.

Amazon QLDB does not support asymmetric keys. For more information, see Using symmetric and asymmetric keys in the Key Management Service Developer Guide.

To specify a customer managed KMS key, you can use its key ID, Amazon Resource Name (ARN), alias name, or alias ARN. When using an alias name, prefix it with "alias/". To specify a key in a different account, you must use the key ARN or alias ARN.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Alias name: alias/ExampleAlias

Alias ARN: arn:aws:kms:us-east-2:111122223333:alias/ExampleAlias

For more information, see Key identifiers (KeyId) in the Key Management Service Developer Guide.

createLedger_tags :: Lens' CreateLedger (Maybe (HashMap Text Text))

The key-value pairs to add as tags to the ledger that you want to create. Tag keys are case sensitive. Tag values are case sensitive and can be null.

createLedger_name :: Lens' CreateLedger Text

The name of the ledger that you want to create. The name must be unique among all of the ledgers in your account in the current Region.

Naming constraints for ledger names are defined in Quotas in Amazon QLDB in the Amazon QLDB Developer Guide.

createLedger_permissionsMode :: Lens' CreateLedger PermissionsMode

The permissions mode to assign to the ledger that you want to create. This parameter can have one of the following values:

ALLOW_ALL: A legacy permissions mode that enables access control with API-level granularity for ledgers.

This mode allows users who have the SendCommand API permission for this ledger to run all PartiQL commands (hence, ALLOW_ALL) on any tables in the specified ledger. This mode disregards any table-level or command-level IAM permissions policies that you create for the ledger.

STANDARD: (Recommended) A permissions mode that enables access control with finer granularity for ledgers, tables, and PartiQL commands.

By default, this mode denies all user requests to run any PartiQL commands on any tables in this ledger. To allow PartiQL commands to run, you must create IAM permissions policies for specific table resources and PartiQL actions, in addition to the SendCommand API permission for the ledger. For information, see Getting started with the standard permissions mode in the Amazon QLDB Developer Guide.

We strongly recommend using the STANDARD permissions mode to maximize the security of your ledger data.

createLedgerResponse_state :: Lens' CreateLedgerResponse (Maybe LedgerState)

The current status of the ledger.

createLedgerResponse_deletionProtection :: Lens' CreateLedgerResponse (Maybe Bool)

The flag that prevents a ledger from being deleted by any user. If not provided on ledger creation, this feature is enabled (true) by default.

If deletion protection is enabled, you must first disable it before you can delete the ledger. You can disable it by calling the UpdateLedger operation to set the flag to false.

createLedgerResponse_kmsKeyArn :: Lens' CreateLedgerResponse (Maybe Text)

The ARN of the customer managed KMS key that the ledger uses for encryption at rest. If this parameter is undefined, the ledger uses an Amazon Web Services owned KMS key for encryption.

createLedgerResponse_arn :: Lens' CreateLedgerResponse (Maybe Text)

The Amazon Resource Name (ARN) for the ledger.

createLedgerResponse_name :: Lens' CreateLedgerResponse (Maybe Text)

The name of the ledger.

createLedgerResponse_creationDateTime :: Lens' CreateLedgerResponse (Maybe UTCTime)

The date and time, in epoch time format, when the ledger was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

createLedgerResponse_permissionsMode :: Lens' CreateLedgerResponse (Maybe PermissionsMode)

The permissions mode of the ledger that you created.

createLedgerResponse_httpStatus :: Lens' CreateLedgerResponse Int

The response's http status code.

ListLedgers

listLedgers_nextToken :: Lens' ListLedgers (Maybe Text)

A pagination token, indicating that you want to retrieve the next page of results. If you received a value for NextToken in the response from a previous ListLedgers call, then you should use that value as input here.

listLedgers_maxResults :: Lens' ListLedgers (Maybe Natural)

The maximum number of results to return in a single ListLedgers request. (The actual number of results returned might be fewer.)

listLedgersResponse_ledgers :: Lens' ListLedgersResponse (Maybe [LedgerSummary])

The array of ledger summaries that are associated with the current account and Region.

listLedgersResponse_nextToken :: Lens' ListLedgersResponse (Maybe Text)

A pagination token, indicating whether there are more results available:

If NextToken is empty, then the last page of results has been processed and there are no more results to be retrieved.

If NextToken is not empty, then there are more results available. To retrieve the next page of results, use the value of NextToken in a subsequent ListLedgers call.

listLedgersResponse_httpStatus :: Lens' ListLedgersResponse Int

The response's http status code.

ListJournalS3Exports

listJournalS3Exports_nextToken :: Lens' ListJournalS3Exports (Maybe Text)

A pagination token, indicating that you want to retrieve the next page of results. If you received a value for NextToken in the response from a previous ListJournalS3Exports call, then you should use that value as input here.

listJournalS3Exports_maxResults :: Lens' ListJournalS3Exports (Maybe Natural)

The maximum number of results to return in a single ListJournalS3Exports request. (The actual number of results returned might be fewer.)

listJournalS3ExportsResponse_journalS3Exports :: Lens' ListJournalS3ExportsResponse (Maybe [JournalS3ExportDescription])

The array of journal export job descriptions for all ledgers that are associated with the current account and Region.

listJournalS3ExportsResponse_nextToken :: Lens' ListJournalS3ExportsResponse (Maybe Text)

If NextToken is empty, then the last page of results has been processed and there are no more results to be retrieved.

If NextToken is not empty, then there are more results available. To retrieve the next page of results, use the value of NextToken in a subsequent ListJournalS3Exports call.

listJournalS3ExportsResponse_httpStatus :: Lens' ListJournalS3ExportsResponse Int

The response's http status code.

UpdateLedgerPermissionsMode

updateLedgerPermissionsMode_name :: Lens' UpdateLedgerPermissionsMode Text

The name of the ledger.

updateLedgerPermissionsMode_permissionsMode :: Lens' UpdateLedgerPermissionsMode PermissionsMode

The permissions mode to assign to the ledger. This parameter can have one of the following values:

ALLOW_ALL: A legacy permissions mode that enables access control with API-level granularity for ledgers.

This mode allows users who have the SendCommand API permission for this ledger to run all PartiQL commands (hence, ALLOW_ALL) on any tables in the specified ledger. This mode disregards any table-level or command-level IAM permissions policies that you create for the ledger.

STANDARD: (Recommended) A permissions mode that enables access control with finer granularity for ledgers, tables, and PartiQL commands.

By default, this mode denies all user requests to run any PartiQL commands on any tables in this ledger. To allow PartiQL commands to run, you must create IAM permissions policies for specific table resources and PartiQL actions, in addition to the SendCommand API permission for the ledger. For information, see Getting started with the standard permissions mode in the Amazon QLDB Developer Guide.

We strongly recommend using the STANDARD permissions mode to maximize the security of your ledger data.

updateLedgerPermissionsModeResponse_arn :: Lens' UpdateLedgerPermissionsModeResponse (Maybe Text)

The Amazon Resource Name (ARN) for the ledger.

updateLedgerPermissionsModeResponse_name :: Lens' UpdateLedgerPermissionsModeResponse (Maybe Text)

The name of the ledger.

updateLedgerPermissionsModeResponse_permissionsMode :: Lens' UpdateLedgerPermissionsModeResponse (Maybe PermissionsMode)

The current permissions mode of the ledger.

updateLedgerPermissionsModeResponse_httpStatus :: Lens' UpdateLedgerPermissionsModeResponse Int

The response's http status code.

GetBlock

getBlock_digestTipAddress :: Lens' GetBlock (Maybe ValueHolder)

The latest block location covered by the digest for which to request a proof. An address is an Amazon Ion structure that has two fields: strandId and sequenceNo.

For example: {strandId:"BlFTjlSXze9BIh1KOszcE3",sequenceNo:49}.

getBlock_name :: Lens' GetBlock Text

The name of the ledger.

getBlock_blockAddress :: Lens' GetBlock ValueHolder

The location of the block that you want to request. An address is an Amazon Ion structure that has two fields: strandId and sequenceNo.

For example: {strandId:"BlFTjlSXze9BIh1KOszcE3",sequenceNo:14}.

getBlockResponse_proof :: Lens' GetBlockResponse (Maybe ValueHolder)

The proof object in Amazon Ion format returned by a GetBlock request. A proof contains the list of hash values required to recalculate the specified digest using a Merkle tree, starting with the specified block.

getBlockResponse_httpStatus :: Lens' GetBlockResponse Int

The response's http status code.

getBlockResponse_block :: Lens' GetBlockResponse ValueHolder

The block data object in Amazon Ion format.

ListJournalS3ExportsForLedger

listJournalS3ExportsForLedger_nextToken :: Lens' ListJournalS3ExportsForLedger (Maybe Text)

A pagination token, indicating that you want to retrieve the next page of results. If you received a value for NextToken in the response from a previous ListJournalS3ExportsForLedger call, then you should use that value as input here.

listJournalS3ExportsForLedger_maxResults :: Lens' ListJournalS3ExportsForLedger (Maybe Natural)

The maximum number of results to return in a single ListJournalS3ExportsForLedger request. (The actual number of results returned might be fewer.)

listJournalS3ExportsForLedger_name :: Lens' ListJournalS3ExportsForLedger Text

The name of the ledger.

listJournalS3ExportsForLedgerResponse_journalS3Exports :: Lens' ListJournalS3ExportsForLedgerResponse (Maybe [JournalS3ExportDescription])

The array of journal export job descriptions that are associated with the specified ledger.

listJournalS3ExportsForLedgerResponse_nextToken :: Lens' ListJournalS3ExportsForLedgerResponse (Maybe Text)

If NextToken is empty, then the last page of results has been processed and there are no more results to be retrieved.

If NextToken is not empty, then there are more results available. To retrieve the next page of results, use the value of NextToken in a subsequent ListJournalS3ExportsForLedger call.

listJournalS3ExportsForLedgerResponse_httpStatus :: Lens' ListJournalS3ExportsForLedgerResponse Int

The response's http status code.

DescribeJournalKinesisStream

describeJournalKinesisStream_ledgerName :: Lens' DescribeJournalKinesisStream Text

The name of the ledger.

describeJournalKinesisStream_streamId :: Lens' DescribeJournalKinesisStream Text

The UUID (represented in Base62-encoded text) of the QLDB journal stream to describe.

describeJournalKinesisStreamResponse_stream :: Lens' DescribeJournalKinesisStreamResponse (Maybe JournalKinesisStreamDescription)

Information about the QLDB journal stream returned by a DescribeJournalS3Export request.

describeJournalKinesisStreamResponse_httpStatus :: Lens' DescribeJournalKinesisStreamResponse Int

The response's http status code.

DescribeJournalS3Export

describeJournalS3Export_name :: Lens' DescribeJournalS3Export Text

The name of the ledger.

describeJournalS3Export_exportId :: Lens' DescribeJournalS3Export Text

The UUID (represented in Base62-encoded text) of the journal export job to describe.

describeJournalS3ExportResponse_httpStatus :: Lens' DescribeJournalS3ExportResponse Int

The response's http status code.

describeJournalS3ExportResponse_exportDescription :: Lens' DescribeJournalS3ExportResponse JournalS3ExportDescription

Information about the journal export job returned by a DescribeJournalS3Export request.

TagResource

tagResource_resourceArn :: Lens' TagResource Text

The Amazon Resource Name (ARN) to which you want to add the tags. For example:

arn:aws:qldb:us-east-1:123456789012:ledger/exampleLedger

tagResource_tags :: Lens' TagResource (HashMap Text Text)

The key-value pairs to add as tags to the specified QLDB resource. Tag keys are case sensitive. If you specify a key that already exists for the resource, your request fails and returns an error. Tag values are case sensitive and can be null.

tagResourceResponse_httpStatus :: Lens' TagResourceResponse Int

The response's http status code.

UntagResource

untagResource_resourceArn :: Lens' UntagResource Text

The Amazon Resource Name (ARN) from which to remove the tags. For example:

arn:aws:qldb:us-east-1:123456789012:ledger/exampleLedger

untagResource_tagKeys :: Lens' UntagResource [Text]

The list of tag keys to remove.

untagResourceResponse_httpStatus :: Lens' UntagResourceResponse Int

The response's http status code.

GetDigest

getDigest_name :: Lens' GetDigest Text

The name of the ledger.

getDigestResponse_httpStatus :: Lens' GetDigestResponse Int

The response's http status code.

getDigestResponse_digest :: Lens' GetDigestResponse ByteString

The 256-bit hash value representing the digest returned by a GetDigest request.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

getDigestResponse_digestTipAddress :: Lens' GetDigestResponse ValueHolder

The latest block location covered by the digest that you requested. An address is an Amazon Ion structure that has two fields: strandId and sequenceNo.

Types

JournalKinesisStreamDescription

journalKinesisStreamDescription_creationTime :: Lens' JournalKinesisStreamDescription (Maybe UTCTime)

The date and time, in epoch time format, when the QLDB journal stream was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

journalKinesisStreamDescription_arn :: Lens' JournalKinesisStreamDescription (Maybe Text)

The Amazon Resource Name (ARN) of the QLDB journal stream.

journalKinesisStreamDescription_inclusiveStartTime :: Lens' JournalKinesisStreamDescription (Maybe UTCTime)

The inclusive start date and time from which to start streaming journal data.

journalKinesisStreamDescription_errorCause :: Lens' JournalKinesisStreamDescription (Maybe ErrorCause)

The error message that describes the reason that a stream has a status of IMPAIRED or FAILED. This is not applicable to streams that have other status values.

journalKinesisStreamDescription_exclusiveEndTime :: Lens' JournalKinesisStreamDescription (Maybe UTCTime)

The exclusive date and time that specifies when the stream ends. If this parameter is undefined, the stream runs indefinitely until you cancel it.

journalKinesisStreamDescription_ledgerName :: Lens' JournalKinesisStreamDescription Text

The name of the ledger.

journalKinesisStreamDescription_roleArn :: Lens' JournalKinesisStreamDescription Text

The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal stream to write data records to a Kinesis Data Streams resource.

journalKinesisStreamDescription_streamId :: Lens' JournalKinesisStreamDescription Text

The UUID (represented in Base62-encoded text) of the QLDB journal stream.

journalKinesisStreamDescription_status :: Lens' JournalKinesisStreamDescription StreamStatus

The current state of the QLDB journal stream.

journalKinesisStreamDescription_kinesisConfiguration :: Lens' JournalKinesisStreamDescription KinesisConfiguration

The configuration settings of the Amazon Kinesis Data Streams destination for a QLDB journal stream.

journalKinesisStreamDescription_streamName :: Lens' JournalKinesisStreamDescription Text

The user-defined name of the QLDB journal stream.

JournalS3ExportDescription

journalS3ExportDescription_ledgerName :: Lens' JournalS3ExportDescription Text

The name of the ledger.

journalS3ExportDescription_exportId :: Lens' JournalS3ExportDescription Text

The UUID (represented in Base62-encoded text) of the journal export job.

journalS3ExportDescription_exportCreationTime :: Lens' JournalS3ExportDescription UTCTime

The date and time, in epoch time format, when the export job was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

journalS3ExportDescription_status :: Lens' JournalS3ExportDescription ExportStatus

The current state of the journal export job.

journalS3ExportDescription_inclusiveStartTime :: Lens' JournalS3ExportDescription UTCTime

The inclusive start date and time for the range of journal contents that are specified in the original export request.

journalS3ExportDescription_exclusiveEndTime :: Lens' JournalS3ExportDescription UTCTime

The exclusive end date and time for the range of journal contents that are specified in the original export request.

journalS3ExportDescription_s3ExportConfiguration :: Lens' JournalS3ExportDescription S3ExportConfiguration

Undocumented member.

journalS3ExportDescription_roleArn :: Lens' JournalS3ExportDescription Text

The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal export job to do the following:

Write objects into your Amazon Simple Storage Service (Amazon S3) bucket.

(Optional) Use your customer master key (CMK) in Key Management Service (KMS) for server-side encryption of your exported data.

KinesisConfiguration

kinesisConfiguration_aggregationEnabled :: Lens' KinesisConfiguration (Maybe Bool)

Enables QLDB to publish multiple data records in a single Kinesis Data Streams record, increasing the number of records sent per API call.

This option is enabled by default. Record aggregation has important implications for processing records and requires de-aggregation in your stream consumer. To learn more, see KPL Key Concepts and Consumer De-aggregation in the Amazon Kinesis Data Streams Developer Guide.

kinesisConfiguration_streamArn :: Lens' KinesisConfiguration Text

The Amazon Resource Name (ARN) of the Kinesis Data Streams resource.

LedgerEncryptionDescription

ledgerEncryptionDescription_inaccessibleKmsKeyDateTime :: Lens' LedgerEncryptionDescription (Maybe UTCTime)

The date and time, in epoch time format, when the KMS key first became inaccessible, in the case of an error. (Epoch time format is the number of seconds that have elapsed since 12:00:00 AM January 1, 1970 UTC.)

This parameter is undefined if the KMS key is accessible.

ledgerEncryptionDescription_kmsKeyArn :: Lens' LedgerEncryptionDescription Text

The Amazon Resource Name (ARN) of the customer managed KMS key that the ledger uses for encryption at rest. If this parameter is undefined, the ledger uses an Amazon Web Services owned KMS key for encryption.

ledgerEncryptionDescription_encryptionStatus :: Lens' LedgerEncryptionDescription EncryptionStatus

The current state of encryption at rest for the ledger. This can be one of the following values:

ENABLED: Encryption is fully enabled using the specified key.

UPDATING: The ledger is actively processing the specified key change.

Key changes in QLDB are asynchronous. The ledger is fully accessible without any performance impact while the key change is being processed. The amount of time it takes to update a key varies depending on the ledger size.

KMS_KEY_INACCESSIBLE: The specified customer managed KMS key is not accessible, and the ledger is impaired. Either the key was disabled or deleted, or the grants on the key were revoked. When a ledger is impaired, it is not accessible and does not accept any read or write requests.

An impaired ledger automatically returns to an active state after you restore the grants on the key, or re-enable the key that was disabled. However, deleting a customer managed KMS key is irreversible. After a key is deleted, you can no longer access the ledgers that are protected with that key, and the data becomes unrecoverable permanently.

LedgerSummary

ledgerSummary_state :: Lens' LedgerSummary (Maybe LedgerState)

The current status of the ledger.

ledgerSummary_name :: Lens' LedgerSummary (Maybe Text)

The name of the ledger.

ledgerSummary_creationDateTime :: Lens' LedgerSummary (Maybe UTCTime)

The date and time, in epoch time format, when the ledger was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

S3EncryptionConfiguration

s3EncryptionConfiguration_kmsKeyArn :: Lens' S3EncryptionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) in Key Management Service (KMS). Amazon S3 does not support asymmetric CMKs.

You must provide a KmsKeyArn if you specify SSE_KMS as the ObjectEncryptionType.

KmsKeyArn is not required if you specify SSE_S3 as the ObjectEncryptionType.

s3EncryptionConfiguration_objectEncryptionType :: Lens' S3EncryptionConfiguration S3ObjectEncryptionType

The Amazon S3 object encryption type.

To learn more about server-side encryption options in Amazon S3, see Protecting Data Using Server-Side Encryption in the Amazon S3 Developer Guide.

S3ExportConfiguration

s3ExportConfiguration_bucket :: Lens' S3ExportConfiguration Text

The Amazon S3 bucket name in which a journal export job writes the journal contents.

The bucket name must comply with the Amazon S3 bucket naming conventions. For more information, see Bucket Restrictions and Limitations in the Amazon S3 Developer Guide.

s3ExportConfiguration_prefix :: Lens' S3ExportConfiguration Text

The prefix for the Amazon S3 bucket in which a journal export job writes the journal contents.

The prefix must comply with Amazon S3 key naming rules and restrictions. For more information, see Object Key and Metadata in the Amazon S3 Developer Guide.

The following are examples of valid Prefix values:

JournalExports-ForMyLedger/Testing/

JournalExports

My:Tests/

s3ExportConfiguration_encryptionConfiguration :: Lens' S3ExportConfiguration S3EncryptionConfiguration

The encryption settings that are used by a journal export job to write data in an Amazon S3 bucket.

ValueHolder

valueHolder_ionText :: Lens' ValueHolder (Maybe Text)

An Amazon Ion plaintext value contained in a ValueHolder structure.