Network.AWS.QLDB.ListJournalKinesisStreamsForLedger

Description

Returns an array of all Amazon QLDB journal stream descriptors for a given ledger. The output of each stream descriptor includes the same details that are returned by DescribeJournalKinesisStream.

This action does not return any expired journal streams. For more information, see Expiration for terminal streams in the Amazon QLDB Developer Guide.

This action returns a maximum of MaxResults items. It is paginated so that you can retrieve all the items by calling ListJournalKinesisStreamsForLedger multiple times.

Creating a Request

data ListJournalKinesisStreamsForLedger

See: newListJournalKinesisStreamsForLedger smart constructor.

Constructors

ListJournalKinesisStreamsForLedger'

Fields

nextToken :: Maybe Text

A pagination token, indicating that you want to retrieve the next page of results. If you received a value for NextToken in the response from a previous ListJournalKinesisStreamsForLedger call, you should use that value as input here.

maxResults :: Maybe Natural

The maximum number of results to return in a single ListJournalKinesisStreamsForLedger request. (The actual number of results returned might be fewer.)

ledgerName :: Text

The name of the ledger.

newListJournalKinesisStreamsForLedger

Arguments

:: Text
$sel:ledgerName:ListJournalKinesisStreamsForLedger'

-> ListJournalKinesisStreamsForLedger

Create a value of ListJournalKinesisStreamsForLedger with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListJournalKinesisStreamsForLedger', listJournalKinesisStreamsForLedger_nextToken - A pagination token, indicating that you want to retrieve the next page of results. If you received a value for NextToken in the response from a previous ListJournalKinesisStreamsForLedger call, you should use that value as input here.

$sel:maxResults:ListJournalKinesisStreamsForLedger', listJournalKinesisStreamsForLedger_maxResults - The maximum number of results to return in a single ListJournalKinesisStreamsForLedger request. (The actual number of results returned might be fewer.)

$sel:ledgerName:ListJournalKinesisStreamsForLedger', listJournalKinesisStreamsForLedger_ledgerName - The name of the ledger.

Request Lenses

listJournalKinesisStreamsForLedger_nextToken :: Lens' ListJournalKinesisStreamsForLedger (Maybe Text)

A pagination token, indicating that you want to retrieve the next page of results. If you received a value for NextToken in the response from a previous ListJournalKinesisStreamsForLedger call, you should use that value as input here.

listJournalKinesisStreamsForLedger_maxResults :: Lens' ListJournalKinesisStreamsForLedger (Maybe Natural)

The maximum number of results to return in a single ListJournalKinesisStreamsForLedger request. (The actual number of results returned might be fewer.)

listJournalKinesisStreamsForLedger_ledgerName :: Lens' ListJournalKinesisStreamsForLedger Text

The name of the ledger.

Destructuring the Response

data ListJournalKinesisStreamsForLedgerResponse

See: newListJournalKinesisStreamsForLedgerResponse smart constructor.

Constructors

ListJournalKinesisStreamsForLedgerResponse'

Fields

nextToken :: Maybe Text

If NextToken is empty, the last page of results has been processed and there are no more results to be retrieved.

If NextToken is not empty, more results are available. To retrieve the next page of results, use the value of NextToken in a subsequent ListJournalKinesisStreamsForLedger call.

streams :: Maybe [JournalKinesisStreamDescription]

The array of QLDB journal stream descriptors that are associated with the given ledger.

httpStatus :: Int

The response's http status code.

newListJournalKinesisStreamsForLedgerResponse

Arguments

:: Int
$sel:httpStatus:ListJournalKinesisStreamsForLedgerResponse'

-> ListJournalKinesisStreamsForLedgerResponse

Create a value of ListJournalKinesisStreamsForLedgerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListJournalKinesisStreamsForLedger', listJournalKinesisStreamsForLedgerResponse_nextToken - - If NextToken is empty, the last page of results has been processed and there are no more results to be retrieved.

If NextToken is not empty, more results are available. To retrieve the next page of results, use the value of NextToken in a subsequent ListJournalKinesisStreamsForLedger call.

$sel:streams:ListJournalKinesisStreamsForLedgerResponse', listJournalKinesisStreamsForLedgerResponse_streams - The array of QLDB journal stream descriptors that are associated with the given ledger.

$sel:httpStatus:ListJournalKinesisStreamsForLedgerResponse', listJournalKinesisStreamsForLedgerResponse_httpStatus - The response's http status code.

Response Lenses

listJournalKinesisStreamsForLedgerResponse_nextToken :: Lens' ListJournalKinesisStreamsForLedgerResponse (Maybe Text)

If NextToken is empty, the last page of results has been processed and there are no more results to be retrieved.

If NextToken is not empty, more results are available. To retrieve the next page of results, use the value of NextToken in a subsequent ListJournalKinesisStreamsForLedger call.

listJournalKinesisStreamsForLedgerResponse_streams :: Lens' ListJournalKinesisStreamsForLedgerResponse (Maybe [JournalKinesisStreamDescription])

The array of QLDB journal stream descriptors that are associated with the given ledger.

listJournalKinesisStreamsForLedgerResponse_httpStatus :: Lens' ListJournalKinesisStreamsForLedgerResponse Int

The response's http status code.