Network.AWS.QLDB.ListJournalS3Exports

Description

Returns an array of journal export job descriptions for all ledgers that are associated with the current account and Region.

This action returns a maximum of MaxResults items, and is paginated so that you can retrieve all the items by calling ListJournalS3Exports multiple times.

This action does not return any expired export jobs. For more information, see Export job expiration in the Amazon QLDB Developer Guide.

Creating a Request

data ListJournalS3Exports

See: newListJournalS3Exports smart constructor.

Constructors

ListJournalS3Exports'

Fields

nextToken :: Maybe Text

A pagination token, indicating that you want to retrieve the next page of results. If you received a value for NextToken in the response from a previous ListJournalS3Exports call, then you should use that value as input here.

maxResults :: Maybe Natural

The maximum number of results to return in a single ListJournalS3Exports request. (The actual number of results returned might be fewer.)

newListJournalS3Exports :: ListJournalS3Exports

Create a value of ListJournalS3Exports with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListJournalS3Exports', listJournalS3Exports_nextToken - A pagination token, indicating that you want to retrieve the next page of results. If you received a value for NextToken in the response from a previous ListJournalS3Exports call, then you should use that value as input here.

$sel:maxResults:ListJournalS3Exports', listJournalS3Exports_maxResults - The maximum number of results to return in a single ListJournalS3Exports request. (The actual number of results returned might be fewer.)

Request Lenses

listJournalS3Exports_nextToken :: Lens' ListJournalS3Exports (Maybe Text)

A pagination token, indicating that you want to retrieve the next page of results. If you received a value for NextToken in the response from a previous ListJournalS3Exports call, then you should use that value as input here.

listJournalS3Exports_maxResults :: Lens' ListJournalS3Exports (Maybe Natural)

The maximum number of results to return in a single ListJournalS3Exports request. (The actual number of results returned might be fewer.)

Destructuring the Response

data ListJournalS3ExportsResponse

See: newListJournalS3ExportsResponse smart constructor.

Constructors

ListJournalS3ExportsResponse'

Fields

journalS3Exports :: Maybe [JournalS3ExportDescription]

The array of journal export job descriptions for all ledgers that are associated with the current account and Region.

nextToken :: Maybe Text

If NextToken is empty, then the last page of results has been processed and there are no more results to be retrieved.

If NextToken is not empty, then there are more results available. To retrieve the next page of results, use the value of NextToken in a subsequent ListJournalS3Exports call.

httpStatus :: Int

The response's http status code.

newListJournalS3ExportsResponse

Arguments

:: Int
$sel:httpStatus:ListJournalS3ExportsResponse'

-> ListJournalS3ExportsResponse

Create a value of ListJournalS3ExportsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:journalS3Exports:ListJournalS3ExportsResponse', listJournalS3ExportsResponse_journalS3Exports - The array of journal export job descriptions for all ledgers that are associated with the current account and Region.

$sel:nextToken:ListJournalS3Exports', listJournalS3ExportsResponse_nextToken - - If NextToken is empty, then the last page of results has been processed and there are no more results to be retrieved.

If NextToken is not empty, then there are more results available. To retrieve the next page of results, use the value of NextToken in a subsequent ListJournalS3Exports call.

$sel:httpStatus:ListJournalS3ExportsResponse', listJournalS3ExportsResponse_httpStatus - The response's http status code.

Response Lenses

listJournalS3ExportsResponse_journalS3Exports :: Lens' ListJournalS3ExportsResponse (Maybe [JournalS3ExportDescription])

The array of journal export job descriptions for all ledgers that are associated with the current account and Region.

listJournalS3ExportsResponse_nextToken :: Lens' ListJournalS3ExportsResponse (Maybe Text)

If NextToken is empty, then the last page of results has been processed and there are no more results to be retrieved.

If NextToken is not empty, then there are more results available. To retrieve the next page of results, use the value of NextToken in a subsequent ListJournalS3Exports call.

listJournalS3ExportsResponse_httpStatus :: Lens' ListJournalS3ExportsResponse Int

The response's http status code.