Network.AWS.QLDB.ListLedgers

Description

Returns an array of ledger summaries that are associated with the current account and Region.

This action returns a maximum of 100 items and is paginated so that you can retrieve all the items by calling ListLedgers multiple times.

Creating a Request

data ListLedgers

See: newListLedgers smart constructor.

Constructors

ListLedgers'

Fields

nextToken :: Maybe Text

A pagination token, indicating that you want to retrieve the next page of results. If you received a value for NextToken in the response from a previous ListLedgers call, then you should use that value as input here.

maxResults :: Maybe Natural

The maximum number of results to return in a single ListLedgers request. (The actual number of results returned might be fewer.)

newListLedgers :: ListLedgers

Create a value of ListLedgers with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListLedgers', listLedgers_nextToken - A pagination token, indicating that you want to retrieve the next page of results. If you received a value for NextToken in the response from a previous ListLedgers call, then you should use that value as input here.

$sel:maxResults:ListLedgers', listLedgers_maxResults - The maximum number of results to return in a single ListLedgers request. (The actual number of results returned might be fewer.)

Request Lenses

listLedgers_nextToken :: Lens' ListLedgers (Maybe Text)

A pagination token, indicating that you want to retrieve the next page of results. If you received a value for NextToken in the response from a previous ListLedgers call, then you should use that value as input here.

listLedgers_maxResults :: Lens' ListLedgers (Maybe Natural)

The maximum number of results to return in a single ListLedgers request. (The actual number of results returned might be fewer.)

Destructuring the Response

data ListLedgersResponse

See: newListLedgersResponse smart constructor.

Constructors

ListLedgersResponse'

Fields

nextToken :: Maybe Text

A pagination token, indicating whether there are more results available:

If NextToken is empty, then the last page of results has been processed and there are no more results to be retrieved.

If NextToken is not empty, then there are more results available. To retrieve the next page of results, use the value of NextToken in a subsequent ListLedgers call.

ledgers :: Maybe [LedgerSummary]

The array of ledger summaries that are associated with the current account and Region.

httpStatus :: Int

The response's http status code.

newListLedgersResponse

Arguments

:: Int
$sel:httpStatus:ListLedgersResponse'

-> ListLedgersResponse

Create a value of ListLedgersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListLedgers', listLedgersResponse_nextToken - A pagination token, indicating whether there are more results available:

If NextToken is empty, then the last page of results has been processed and there are no more results to be retrieved.

If NextToken is not empty, then there are more results available. To retrieve the next page of results, use the value of NextToken in a subsequent ListLedgers call.

$sel:ledgers:ListLedgersResponse', listLedgersResponse_ledgers - The array of ledger summaries that are associated with the current account and Region.

$sel:httpStatus:ListLedgersResponse', listLedgersResponse_httpStatus - The response's http status code.

Response Lenses

listLedgersResponse_nextToken :: Lens' ListLedgersResponse (Maybe Text)

A pagination token, indicating whether there are more results available:

If NextToken is empty, then the last page of results has been processed and there are no more results to be retrieved.

If NextToken is not empty, then there are more results available. To retrieve the next page of results, use the value of NextToken in a subsequent ListLedgers call.

listLedgersResponse_ledgers :: Lens' ListLedgersResponse (Maybe [LedgerSummary])

The array of ledger summaries that are associated with the current account and Region.

listLedgersResponse_httpStatus :: Lens' ListLedgersResponse Int

The response's http status code.