Network.AWS.QLDB.StreamJournalToKinesis

Description

Creates a journal stream for a given Amazon QLDB ledger. The stream captures every document revision that is committed to the ledger's journal and delivers the data to a specified Amazon Kinesis Data Streams resource.

Creating a Request

data StreamJournalToKinesis

See: newStreamJournalToKinesis smart constructor.

Constructors

StreamJournalToKinesis'

Fields

exclusiveEndTime :: Maybe POSIX

The exclusive date and time that specifies when the stream ends. If you don't define this parameter, the stream runs indefinitely until you cancel it.

The ExclusiveEndTime must be in ISO 8601 date and time format and in Universal Coordinated Time (UTC). For example: 2019-06-13T21:36:34Z.

tags :: Maybe (HashMap Text Text)

The key-value pairs to add as tags to the stream that you want to create. Tag keys are case sensitive. Tag values are case sensitive and can be null.

ledgerName :: Text

The name of the ledger.

roleArn :: Text

The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal stream to write data records to a Kinesis Data Streams resource.

inclusiveStartTime :: POSIX

The inclusive start date and time from which to start streaming journal data. This parameter must be in ISO 8601 date and time format and in Universal Coordinated Time (UTC). For example: 2019-06-13T21:36:34Z.

The InclusiveStartTime cannot be in the future and must be before ExclusiveEndTime.

If you provide an InclusiveStartTime that is before the ledger's CreationDateTime, QLDB effectively defaults it to the ledger's CreationDateTime.

kinesisConfiguration :: KinesisConfiguration

The configuration settings of the Kinesis Data Streams destination for your stream request.

streamName :: Text

The name that you want to assign to the QLDB journal stream. User-defined names can help identify and indicate the purpose of a stream.

Your stream name must be unique among other active streams for a given ledger. Stream names have the same naming constraints as ledger names, as defined in Quotas in Amazon QLDB in the Amazon QLDB Developer Guide.

newStreamJournalToKinesis

Arguments

:: Text
$sel:ledgerName:StreamJournalToKinesis'

-> Text
$sel:roleArn:StreamJournalToKinesis'

-> UTCTime
$sel:inclusiveStartTime:StreamJournalToKinesis'

-> KinesisConfiguration
$sel:kinesisConfiguration:StreamJournalToKinesis'

-> Text
$sel:streamName:StreamJournalToKinesis'

-> StreamJournalToKinesis

Create a value of StreamJournalToKinesis with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:exclusiveEndTime:StreamJournalToKinesis', streamJournalToKinesis_exclusiveEndTime - The exclusive date and time that specifies when the stream ends. If you don't define this parameter, the stream runs indefinitely until you cancel it.

The ExclusiveEndTime must be in ISO 8601 date and time format and in Universal Coordinated Time (UTC). For example: 2019-06-13T21:36:34Z.

$sel:tags:StreamJournalToKinesis', streamJournalToKinesis_tags - The key-value pairs to add as tags to the stream that you want to create. Tag keys are case sensitive. Tag values are case sensitive and can be null.

$sel:ledgerName:StreamJournalToKinesis', streamJournalToKinesis_ledgerName - The name of the ledger.

$sel:roleArn:StreamJournalToKinesis', streamJournalToKinesis_roleArn - The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal stream to write data records to a Kinesis Data Streams resource.

$sel:inclusiveStartTime:StreamJournalToKinesis', streamJournalToKinesis_inclusiveStartTime - The inclusive start date and time from which to start streaming journal data. This parameter must be in ISO 8601 date and time format and in Universal Coordinated Time (UTC). For example: 2019-06-13T21:36:34Z.

The InclusiveStartTime cannot be in the future and must be before ExclusiveEndTime.

If you provide an InclusiveStartTime that is before the ledger's CreationDateTime, QLDB effectively defaults it to the ledger's CreationDateTime.

$sel:kinesisConfiguration:StreamJournalToKinesis', streamJournalToKinesis_kinesisConfiguration - The configuration settings of the Kinesis Data Streams destination for your stream request.

$sel:streamName:StreamJournalToKinesis', streamJournalToKinesis_streamName - The name that you want to assign to the QLDB journal stream. User-defined names can help identify and indicate the purpose of a stream.

Your stream name must be unique among other active streams for a given ledger. Stream names have the same naming constraints as ledger names, as defined in Quotas in Amazon QLDB in the Amazon QLDB Developer Guide.

Request Lenses

streamJournalToKinesis_exclusiveEndTime :: Lens' StreamJournalToKinesis (Maybe UTCTime)

The exclusive date and time that specifies when the stream ends. If you don't define this parameter, the stream runs indefinitely until you cancel it.

The ExclusiveEndTime must be in ISO 8601 date and time format and in Universal Coordinated Time (UTC). For example: 2019-06-13T21:36:34Z.

streamJournalToKinesis_tags :: Lens' StreamJournalToKinesis (Maybe (HashMap Text Text))

The key-value pairs to add as tags to the stream that you want to create. Tag keys are case sensitive. Tag values are case sensitive and can be null.

streamJournalToKinesis_ledgerName :: Lens' StreamJournalToKinesis Text

The name of the ledger.

streamJournalToKinesis_roleArn :: Lens' StreamJournalToKinesis Text

The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal stream to write data records to a Kinesis Data Streams resource.

streamJournalToKinesis_inclusiveStartTime :: Lens' StreamJournalToKinesis UTCTime

The inclusive start date and time from which to start streaming journal data. This parameter must be in ISO 8601 date and time format and in Universal Coordinated Time (UTC). For example: 2019-06-13T21:36:34Z.

The InclusiveStartTime cannot be in the future and must be before ExclusiveEndTime.

If you provide an InclusiveStartTime that is before the ledger's CreationDateTime, QLDB effectively defaults it to the ledger's CreationDateTime.

streamJournalToKinesis_kinesisConfiguration :: Lens' StreamJournalToKinesis KinesisConfiguration

The configuration settings of the Kinesis Data Streams destination for your stream request.

streamJournalToKinesis_streamName :: Lens' StreamJournalToKinesis Text

The name that you want to assign to the QLDB journal stream. User-defined names can help identify and indicate the purpose of a stream.

Your stream name must be unique among other active streams for a given ledger. Stream names have the same naming constraints as ledger names, as defined in Quotas in Amazon QLDB in the Amazon QLDB Developer Guide.

Destructuring the Response

data StreamJournalToKinesisResponse

See: newStreamJournalToKinesisResponse smart constructor.

Constructors

StreamJournalToKinesisResponse'

Fields

streamId :: Maybe Text

The UUID (represented in Base62-encoded text) that QLDB assigns to each QLDB journal stream.

httpStatus :: Int

The response's http status code.

newStreamJournalToKinesisResponse

Arguments

:: Int
$sel:httpStatus:StreamJournalToKinesisResponse'

-> StreamJournalToKinesisResponse

Create a value of StreamJournalToKinesisResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:streamId:StreamJournalToKinesisResponse', streamJournalToKinesisResponse_streamId - The UUID (represented in Base62-encoded text) that QLDB assigns to each QLDB journal stream.

$sel:httpStatus:StreamJournalToKinesisResponse', streamJournalToKinesisResponse_httpStatus - The response's http status code.

Response Lenses

streamJournalToKinesisResponse_streamId :: Lens' StreamJournalToKinesisResponse (Maybe Text)

The UUID (represented in Base62-encoded text) that QLDB assigns to each QLDB journal stream.

streamJournalToKinesisResponse_httpStatus :: Lens' StreamJournalToKinesisResponse Int

The response's http status code.