Network.AWS.QLDB.Types

Description

Service Configuration

defaultService :: Service

API version 2019-01-02 of the Amazon QLDB SDK configuration.

Errors

_ResourceAlreadyExistsException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource already exists.

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

One or more parameters in the request aren't valid.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

You have reached the limit on the maximum number of resources allowed.

_ResourceInUseException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource can't be modified at this time.

_ResourceNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource doesn't exist.

_ResourcePreconditionNotMetException :: AsError a => Getting (First ServiceError) a ServiceError

The operation failed because a condition wasn't satisfied in advance.

EncryptionStatus

newtype EncryptionStatus

Constructors

EncryptionStatus'

Fields

fromEncryptionStatus :: Text

Bundled Patterns

pattern EncryptionStatus_ENABLED :: EncryptionStatus

pattern EncryptionStatus_KMS_KEY_INACCESSIBLE :: EncryptionStatus

pattern EncryptionStatus_UPDATING :: EncryptionStatus

ErrorCause

newtype ErrorCause

Constructors

ErrorCause'

Fields

fromErrorCause :: Text

Bundled Patterns

pattern ErrorCause_IAM_PERMISSION_REVOKED :: ErrorCause

pattern ErrorCause_KINESIS_STREAM_NOT_FOUND :: ErrorCause

ExportStatus

newtype ExportStatus

Constructors

ExportStatus'

Fields

fromExportStatus :: Text

Bundled Patterns

pattern ExportStatus_CANCELLED :: ExportStatus

pattern ExportStatus_COMPLETED :: ExportStatus

pattern ExportStatus_IN_PROGRESS :: ExportStatus

LedgerState

newtype LedgerState

Constructors

LedgerState'

Fields

fromLedgerState :: Text

Bundled Patterns

pattern LedgerState_ACTIVE :: LedgerState

pattern LedgerState_CREATING :: LedgerState

pattern LedgerState_DELETED :: LedgerState

pattern LedgerState_DELETING :: LedgerState

PermissionsMode

newtype PermissionsMode

Constructors

PermissionsMode'

Fields

fromPermissionsMode :: Text

Bundled Patterns

pattern PermissionsMode_ALLOW_ALL :: PermissionsMode

pattern PermissionsMode_STANDARD :: PermissionsMode

S3ObjectEncryptionType

newtype S3ObjectEncryptionType

Constructors

S3ObjectEncryptionType'

Fields

fromS3ObjectEncryptionType :: Text

Bundled Patterns

pattern S3ObjectEncryptionType_NO_ENCRYPTION :: S3ObjectEncryptionType

pattern S3ObjectEncryptionType_SSE_KMS :: S3ObjectEncryptionType

pattern S3ObjectEncryptionType_SSE_S3 :: S3ObjectEncryptionType

StreamStatus

newtype StreamStatus

Constructors

StreamStatus'

Fields

fromStreamStatus :: Text

Bundled Patterns

pattern StreamStatus_ACTIVE :: StreamStatus

pattern StreamStatus_CANCELED :: StreamStatus

pattern StreamStatus_COMPLETED :: StreamStatus

pattern StreamStatus_FAILED :: StreamStatus

pattern StreamStatus_IMPAIRED :: StreamStatus

JournalKinesisStreamDescription

data JournalKinesisStreamDescription

Information about an Amazon QLDB journal stream, including the Amazon Resource Name (ARN), stream name, creation time, current status, and the parameters of the original stream creation request.

See: newJournalKinesisStreamDescription smart constructor.

Constructors

JournalKinesisStreamDescription'

Fields

creationTime :: Maybe POSIX

The date and time, in epoch time format, when the QLDB journal stream was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

errorCause :: Maybe ErrorCause

The error message that describes the reason that a stream has a status of IMPAIRED or FAILED. This is not applicable to streams that have other status values.

inclusiveStartTime :: Maybe POSIX

The inclusive start date and time from which to start streaming journal data.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the QLDB journal stream.

exclusiveEndTime :: Maybe POSIX

The exclusive date and time that specifies when the stream ends. If this parameter is undefined, the stream runs indefinitely until you cancel it.

ledgerName :: Text

The name of the ledger.

roleArn :: Text

The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal stream to write data records to a Kinesis Data Streams resource.

streamId :: Text

The UUID (represented in Base62-encoded text) of the QLDB journal stream.

status :: StreamStatus

The current state of the QLDB journal stream.

kinesisConfiguration :: KinesisConfiguration

The configuration settings of the Amazon Kinesis Data Streams destination for a QLDB journal stream.

streamName :: Text

The user-defined name of the QLDB journal stream.

newJournalKinesisStreamDescription

Arguments

:: Text
$sel:ledgerName:JournalKinesisStreamDescription'

-> Text
$sel:roleArn:JournalKinesisStreamDescription'

-> Text
$sel:streamId:JournalKinesisStreamDescription'

-> StreamStatus
$sel:status:JournalKinesisStreamDescription'

-> KinesisConfiguration
$sel:kinesisConfiguration:JournalKinesisStreamDescription'

-> Text
$sel:streamName:JournalKinesisStreamDescription'

-> JournalKinesisStreamDescription

Create a value of JournalKinesisStreamDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:JournalKinesisStreamDescription', journalKinesisStreamDescription_creationTime - The date and time, in epoch time format, when the QLDB journal stream was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

$sel:errorCause:JournalKinesisStreamDescription', journalKinesisStreamDescription_errorCause - The error message that describes the reason that a stream has a status of IMPAIRED or FAILED. This is not applicable to streams that have other status values.

$sel:inclusiveStartTime:JournalKinesisStreamDescription', journalKinesisStreamDescription_inclusiveStartTime - The inclusive start date and time from which to start streaming journal data.

$sel:arn:JournalKinesisStreamDescription', journalKinesisStreamDescription_arn - The Amazon Resource Name (ARN) of the QLDB journal stream.

$sel:exclusiveEndTime:JournalKinesisStreamDescription', journalKinesisStreamDescription_exclusiveEndTime - The exclusive date and time that specifies when the stream ends. If this parameter is undefined, the stream runs indefinitely until you cancel it.

$sel:ledgerName:JournalKinesisStreamDescription', journalKinesisStreamDescription_ledgerName - The name of the ledger.

$sel:roleArn:JournalKinesisStreamDescription', journalKinesisStreamDescription_roleArn - The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal stream to write data records to a Kinesis Data Streams resource.

$sel:streamId:JournalKinesisStreamDescription', journalKinesisStreamDescription_streamId - The UUID (represented in Base62-encoded text) of the QLDB journal stream.

$sel:status:JournalKinesisStreamDescription', journalKinesisStreamDescription_status - The current state of the QLDB journal stream.

$sel:kinesisConfiguration:JournalKinesisStreamDescription', journalKinesisStreamDescription_kinesisConfiguration - The configuration settings of the Amazon Kinesis Data Streams destination for a QLDB journal stream.

$sel:streamName:JournalKinesisStreamDescription', journalKinesisStreamDescription_streamName - The user-defined name of the QLDB journal stream.

journalKinesisStreamDescription_creationTime :: Lens' JournalKinesisStreamDescription (Maybe UTCTime)

The date and time, in epoch time format, when the QLDB journal stream was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

journalKinesisStreamDescription_errorCause :: Lens' JournalKinesisStreamDescription (Maybe ErrorCause)

The error message that describes the reason that a stream has a status of IMPAIRED or FAILED. This is not applicable to streams that have other status values.

journalKinesisStreamDescription_inclusiveStartTime :: Lens' JournalKinesisStreamDescription (Maybe UTCTime)

The inclusive start date and time from which to start streaming journal data.

journalKinesisStreamDescription_arn :: Lens' JournalKinesisStreamDescription (Maybe Text)

The Amazon Resource Name (ARN) of the QLDB journal stream.

journalKinesisStreamDescription_exclusiveEndTime :: Lens' JournalKinesisStreamDescription (Maybe UTCTime)

The exclusive date and time that specifies when the stream ends. If this parameter is undefined, the stream runs indefinitely until you cancel it.

journalKinesisStreamDescription_ledgerName :: Lens' JournalKinesisStreamDescription Text

The name of the ledger.

journalKinesisStreamDescription_roleArn :: Lens' JournalKinesisStreamDescription Text

The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal stream to write data records to a Kinesis Data Streams resource.

journalKinesisStreamDescription_streamId :: Lens' JournalKinesisStreamDescription Text

The UUID (represented in Base62-encoded text) of the QLDB journal stream.

journalKinesisStreamDescription_status :: Lens' JournalKinesisStreamDescription StreamStatus

The current state of the QLDB journal stream.

journalKinesisStreamDescription_kinesisConfiguration :: Lens' JournalKinesisStreamDescription KinesisConfiguration

The configuration settings of the Amazon Kinesis Data Streams destination for a QLDB journal stream.

journalKinesisStreamDescription_streamName :: Lens' JournalKinesisStreamDescription Text

The user-defined name of the QLDB journal stream.

JournalS3ExportDescription

data JournalS3ExportDescription

Information about a journal export job, including the ledger name, export ID, creation time, current status, and the parameters of the original export creation request.

See: newJournalS3ExportDescription smart constructor.

Constructors

JournalS3ExportDescription'

Fields

ledgerName :: Text

The name of the ledger.

exportId :: Text

The UUID (represented in Base62-encoded text) of the journal export job.

exportCreationTime :: POSIX

The date and time, in epoch time format, when the export job was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

status :: ExportStatus

The current state of the journal export job.

inclusiveStartTime :: POSIX

The inclusive start date and time for the range of journal contents that are specified in the original export request.

exclusiveEndTime :: POSIX

The exclusive end date and time for the range of journal contents that are specified in the original export request.

s3ExportConfiguration :: S3ExportConfiguration

roleArn :: Text

The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal export job to do the following:

Write objects into your Amazon Simple Storage Service (Amazon S3) bucket.

(Optional) Use your customer master key (CMK) in Key Management Service (KMS) for server-side encryption of your exported data.

newJournalS3ExportDescription

Arguments

:: Text
$sel:ledgerName:JournalS3ExportDescription'

-> Text
$sel:exportId:JournalS3ExportDescription'

-> UTCTime
$sel:exportCreationTime:JournalS3ExportDescription'

-> ExportStatus
$sel:status:JournalS3ExportDescription'

-> UTCTime
$sel:inclusiveStartTime:JournalS3ExportDescription'

-> UTCTime
$sel:exclusiveEndTime:JournalS3ExportDescription'

-> S3ExportConfiguration
$sel:s3ExportConfiguration:JournalS3ExportDescription'

-> Text
$sel:roleArn:JournalS3ExportDescription'

-> JournalS3ExportDescription

Create a value of JournalS3ExportDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ledgerName:JournalS3ExportDescription', journalS3ExportDescription_ledgerName - The name of the ledger.

$sel:exportId:JournalS3ExportDescription', journalS3ExportDescription_exportId - The UUID (represented in Base62-encoded text) of the journal export job.

$sel:exportCreationTime:JournalS3ExportDescription', journalS3ExportDescription_exportCreationTime - The date and time, in epoch time format, when the export job was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

$sel:status:JournalS3ExportDescription', journalS3ExportDescription_status - The current state of the journal export job.

$sel:inclusiveStartTime:JournalS3ExportDescription', journalS3ExportDescription_inclusiveStartTime - The inclusive start date and time for the range of journal contents that are specified in the original export request.

$sel:exclusiveEndTime:JournalS3ExportDescription', journalS3ExportDescription_exclusiveEndTime - The exclusive end date and time for the range of journal contents that are specified in the original export request.

$sel:s3ExportConfiguration:JournalS3ExportDescription', journalS3ExportDescription_s3ExportConfiguration - Undocumented member.

$sel:roleArn:JournalS3ExportDescription', journalS3ExportDescription_roleArn - The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal export job to do the following:

Write objects into your Amazon Simple Storage Service (Amazon S3) bucket.

(Optional) Use your customer master key (CMK) in Key Management Service (KMS) for server-side encryption of your exported data.

journalS3ExportDescription_ledgerName :: Lens' JournalS3ExportDescription Text

The name of the ledger.

journalS3ExportDescription_exportId :: Lens' JournalS3ExportDescription Text

The UUID (represented in Base62-encoded text) of the journal export job.

journalS3ExportDescription_exportCreationTime :: Lens' JournalS3ExportDescription UTCTime

The date and time, in epoch time format, when the export job was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

journalS3ExportDescription_status :: Lens' JournalS3ExportDescription ExportStatus

The current state of the journal export job.

journalS3ExportDescription_inclusiveStartTime :: Lens' JournalS3ExportDescription UTCTime

The inclusive start date and time for the range of journal contents that are specified in the original export request.

journalS3ExportDescription_exclusiveEndTime :: Lens' JournalS3ExportDescription UTCTime

The exclusive end date and time for the range of journal contents that are specified in the original export request.

journalS3ExportDescription_s3ExportConfiguration :: Lens' JournalS3ExportDescription S3ExportConfiguration

Undocumented member.

journalS3ExportDescription_roleArn :: Lens' JournalS3ExportDescription Text

The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal export job to do the following:

Write objects into your Amazon Simple Storage Service (Amazon S3) bucket.

(Optional) Use your customer master key (CMK) in Key Management Service (KMS) for server-side encryption of your exported data.

KinesisConfiguration

data KinesisConfiguration

The configuration settings of the Amazon Kinesis Data Streams destination for an Amazon QLDB journal stream.

See: newKinesisConfiguration smart constructor.

Constructors

KinesisConfiguration'

Fields

aggregationEnabled :: Maybe Bool

Enables QLDB to publish multiple data records in a single Kinesis Data Streams record, increasing the number of records sent per API call.

This option is enabled by default. Record aggregation has important implications for processing records and requires de-aggregation in your stream consumer. To learn more, see KPL Key Concepts and Consumer De-aggregation in the Amazon Kinesis Data Streams Developer Guide.

streamArn :: Text

The Amazon Resource Name (ARN) of the Kinesis Data Streams resource.

newKinesisConfiguration

Arguments

:: Text
$sel:streamArn:KinesisConfiguration'

-> KinesisConfiguration

Create a value of KinesisConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aggregationEnabled:KinesisConfiguration', kinesisConfiguration_aggregationEnabled - Enables QLDB to publish multiple data records in a single Kinesis Data Streams record, increasing the number of records sent per API call.

This option is enabled by default. Record aggregation has important implications for processing records and requires de-aggregation in your stream consumer. To learn more, see KPL Key Concepts and Consumer De-aggregation in the Amazon Kinesis Data Streams Developer Guide.

$sel:streamArn:KinesisConfiguration', kinesisConfiguration_streamArn - The Amazon Resource Name (ARN) of the Kinesis Data Streams resource.

kinesisConfiguration_aggregationEnabled :: Lens' KinesisConfiguration (Maybe Bool)

Enables QLDB to publish multiple data records in a single Kinesis Data Streams record, increasing the number of records sent per API call.

This option is enabled by default. Record aggregation has important implications for processing records and requires de-aggregation in your stream consumer. To learn more, see KPL Key Concepts and Consumer De-aggregation in the Amazon Kinesis Data Streams Developer Guide.

kinesisConfiguration_streamArn :: Lens' KinesisConfiguration Text

The Amazon Resource Name (ARN) of the Kinesis Data Streams resource.

LedgerEncryptionDescription

data LedgerEncryptionDescription

Information about the encryption of data at rest in an Amazon QLDB ledger. This includes the current status, the key in Key Management Service (KMS), and when the key became inaccessible (in the case of an error).

For more information, see Encryption at rest in the Amazon QLDB Developer Guide.

See: newLedgerEncryptionDescription smart constructor.

Constructors

LedgerEncryptionDescription'

Fields

inaccessibleKmsKeyDateTime :: Maybe POSIX

The date and time, in epoch time format, when the KMS key first became inaccessible, in the case of an error. (Epoch time format is the number of seconds that have elapsed since 12:00:00 AM January 1, 1970 UTC.)

This parameter is undefined if the KMS key is accessible.

kmsKeyArn :: Text

The Amazon Resource Name (ARN) of the customer managed KMS key that the ledger uses for encryption at rest. If this parameter is undefined, the ledger uses an Amazon Web Services owned KMS key for encryption.

encryptionStatus :: EncryptionStatus

The current state of encryption at rest for the ledger. This can be one of the following values:

ENABLED: Encryption is fully enabled using the specified key.

UPDATING: The ledger is actively processing the specified key change.

Key changes in QLDB are asynchronous. The ledger is fully accessible without any performance impact while the key change is being processed. The amount of time it takes to update a key varies depending on the ledger size.

KMS_KEY_INACCESSIBLE: The specified customer managed KMS key is not accessible, and the ledger is impaired. Either the key was disabled or deleted, or the grants on the key were revoked. When a ledger is impaired, it is not accessible and does not accept any read or write requests.

An impaired ledger automatically returns to an active state after you restore the grants on the key, or re-enable the key that was disabled. However, deleting a customer managed KMS key is irreversible. After a key is deleted, you can no longer access the ledgers that are protected with that key, and the data becomes unrecoverable permanently.

newLedgerEncryptionDescription

Arguments

:: Text
$sel:kmsKeyArn:LedgerEncryptionDescription'

-> EncryptionStatus
$sel:encryptionStatus:LedgerEncryptionDescription'

-> LedgerEncryptionDescription

Create a value of LedgerEncryptionDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:inaccessibleKmsKeyDateTime:LedgerEncryptionDescription', ledgerEncryptionDescription_inaccessibleKmsKeyDateTime - The date and time, in epoch time format, when the KMS key first became inaccessible, in the case of an error. (Epoch time format is the number of seconds that have elapsed since 12:00:00 AM January 1, 1970 UTC.)

This parameter is undefined if the KMS key is accessible.

$sel:kmsKeyArn:LedgerEncryptionDescription', ledgerEncryptionDescription_kmsKeyArn - The Amazon Resource Name (ARN) of the customer managed KMS key that the ledger uses for encryption at rest. If this parameter is undefined, the ledger uses an Amazon Web Services owned KMS key for encryption.

$sel:encryptionStatus:LedgerEncryptionDescription', ledgerEncryptionDescription_encryptionStatus - The current state of encryption at rest for the ledger. This can be one of the following values:

ENABLED: Encryption is fully enabled using the specified key.

UPDATING: The ledger is actively processing the specified key change.

Key changes in QLDB are asynchronous. The ledger is fully accessible without any performance impact while the key change is being processed. The amount of time it takes to update a key varies depending on the ledger size.

KMS_KEY_INACCESSIBLE: The specified customer managed KMS key is not accessible, and the ledger is impaired. Either the key was disabled or deleted, or the grants on the key were revoked. When a ledger is impaired, it is not accessible and does not accept any read or write requests.

An impaired ledger automatically returns to an active state after you restore the grants on the key, or re-enable the key that was disabled. However, deleting a customer managed KMS key is irreversible. After a key is deleted, you can no longer access the ledgers that are protected with that key, and the data becomes unrecoverable permanently.

ledgerEncryptionDescription_inaccessibleKmsKeyDateTime :: Lens' LedgerEncryptionDescription (Maybe UTCTime)

The date and time, in epoch time format, when the KMS key first became inaccessible, in the case of an error. (Epoch time format is the number of seconds that have elapsed since 12:00:00 AM January 1, 1970 UTC.)

This parameter is undefined if the KMS key is accessible.

ledgerEncryptionDescription_kmsKeyArn :: Lens' LedgerEncryptionDescription Text

The Amazon Resource Name (ARN) of the customer managed KMS key that the ledger uses for encryption at rest. If this parameter is undefined, the ledger uses an Amazon Web Services owned KMS key for encryption.

ledgerEncryptionDescription_encryptionStatus :: Lens' LedgerEncryptionDescription EncryptionStatus

The current state of encryption at rest for the ledger. This can be one of the following values:

ENABLED: Encryption is fully enabled using the specified key.

UPDATING: The ledger is actively processing the specified key change.

Key changes in QLDB are asynchronous. The ledger is fully accessible without any performance impact while the key change is being processed. The amount of time it takes to update a key varies depending on the ledger size.

KMS_KEY_INACCESSIBLE: The specified customer managed KMS key is not accessible, and the ledger is impaired. Either the key was disabled or deleted, or the grants on the key were revoked. When a ledger is impaired, it is not accessible and does not accept any read or write requests.

An impaired ledger automatically returns to an active state after you restore the grants on the key, or re-enable the key that was disabled. However, deleting a customer managed KMS key is irreversible. After a key is deleted, you can no longer access the ledgers that are protected with that key, and the data becomes unrecoverable permanently.

LedgerSummary

data LedgerSummary

Information about a ledger, including its name, state, and when it was created.

See: newLedgerSummary smart constructor.

Constructors

LedgerSummary'

Fields

name :: Maybe Text

The name of the ledger.

state :: Maybe LedgerState

The current status of the ledger.

creationDateTime :: Maybe POSIX

The date and time, in epoch time format, when the ledger was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

newLedgerSummary :: LedgerSummary

Create a value of LedgerSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:LedgerSummary', ledgerSummary_name - The name of the ledger.

$sel:state:LedgerSummary', ledgerSummary_state - The current status of the ledger.

$sel:creationDateTime:LedgerSummary', ledgerSummary_creationDateTime - The date and time, in epoch time format, when the ledger was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

ledgerSummary_name :: Lens' LedgerSummary (Maybe Text)

The name of the ledger.

ledgerSummary_state :: Lens' LedgerSummary (Maybe LedgerState)

The current status of the ledger.

ledgerSummary_creationDateTime :: Lens' LedgerSummary (Maybe UTCTime)

The date and time, in epoch time format, when the ledger was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

S3EncryptionConfiguration

data S3EncryptionConfiguration

The encryption settings that are used by a journal export job to write data in an Amazon Simple Storage Service (Amazon S3) bucket.

See: newS3EncryptionConfiguration smart constructor.

Constructors

S3EncryptionConfiguration'

Fields

kmsKeyArn :: Maybe Text

The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) in Key Management Service (KMS). Amazon S3 does not support asymmetric CMKs.

You must provide a KmsKeyArn if you specify SSE_KMS as the ObjectEncryptionType.

KmsKeyArn is not required if you specify SSE_S3 as the ObjectEncryptionType.

objectEncryptionType :: S3ObjectEncryptionType

The Amazon S3 object encryption type.

To learn more about server-side encryption options in Amazon S3, see Protecting Data Using Server-Side Encryption in the Amazon S3 Developer Guide.

newS3EncryptionConfiguration

Arguments

:: S3ObjectEncryptionType
$sel:objectEncryptionType:S3EncryptionConfiguration'

-> S3EncryptionConfiguration

Create a value of S3EncryptionConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:kmsKeyArn:S3EncryptionConfiguration', s3EncryptionConfiguration_kmsKeyArn - The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) in Key Management Service (KMS). Amazon S3 does not support asymmetric CMKs.

You must provide a KmsKeyArn if you specify SSE_KMS as the ObjectEncryptionType.

KmsKeyArn is not required if you specify SSE_S3 as the ObjectEncryptionType.

$sel:objectEncryptionType:S3EncryptionConfiguration', s3EncryptionConfiguration_objectEncryptionType - The Amazon S3 object encryption type.

To learn more about server-side encryption options in Amazon S3, see Protecting Data Using Server-Side Encryption in the Amazon S3 Developer Guide.

s3EncryptionConfiguration_kmsKeyArn :: Lens' S3EncryptionConfiguration (Maybe Text)

The Amazon Resource Name (ARN) of a symmetric customer master key (CMK) in Key Management Service (KMS). Amazon S3 does not support asymmetric CMKs.

You must provide a KmsKeyArn if you specify SSE_KMS as the ObjectEncryptionType.

KmsKeyArn is not required if you specify SSE_S3 as the ObjectEncryptionType.

s3EncryptionConfiguration_objectEncryptionType :: Lens' S3EncryptionConfiguration S3ObjectEncryptionType

The Amazon S3 object encryption type.

To learn more about server-side encryption options in Amazon S3, see Protecting Data Using Server-Side Encryption in the Amazon S3 Developer Guide.

S3ExportConfiguration

data S3ExportConfiguration

The Amazon Simple Storage Service (Amazon S3) bucket location in which a journal export job writes the journal contents.

See: newS3ExportConfiguration smart constructor.

Constructors

S3ExportConfiguration'

Fields

bucket :: Text

The Amazon S3 bucket name in which a journal export job writes the journal contents.

The bucket name must comply with the Amazon S3 bucket naming conventions. For more information, see Bucket Restrictions and Limitations in the Amazon S3 Developer Guide.

prefix :: Text

The prefix for the Amazon S3 bucket in which a journal export job writes the journal contents.

The prefix must comply with Amazon S3 key naming rules and restrictions. For more information, see Object Key and Metadata in the Amazon S3 Developer Guide.

The following are examples of valid Prefix values:

JournalExports-ForMyLedger/Testing/

JournalExports

My:Tests/

encryptionConfiguration :: S3EncryptionConfiguration

The encryption settings that are used by a journal export job to write data in an Amazon S3 bucket.

newS3ExportConfiguration

Arguments

:: Text
$sel:bucket:S3ExportConfiguration'

-> Text
$sel:prefix:S3ExportConfiguration'

-> S3EncryptionConfiguration
$sel:encryptionConfiguration:S3ExportConfiguration'

-> S3ExportConfiguration

Create a value of S3ExportConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:bucket:S3ExportConfiguration', s3ExportConfiguration_bucket - The Amazon S3 bucket name in which a journal export job writes the journal contents.

The bucket name must comply with the Amazon S3 bucket naming conventions. For more information, see Bucket Restrictions and Limitations in the Amazon S3 Developer Guide.

$sel:prefix:S3ExportConfiguration', s3ExportConfiguration_prefix - The prefix for the Amazon S3 bucket in which a journal export job writes the journal contents.

The prefix must comply with Amazon S3 key naming rules and restrictions. For more information, see Object Key and Metadata in the Amazon S3 Developer Guide.

The following are examples of valid Prefix values:

JournalExports-ForMyLedger/Testing/

JournalExports

My:Tests/

$sel:encryptionConfiguration:S3ExportConfiguration', s3ExportConfiguration_encryptionConfiguration - The encryption settings that are used by a journal export job to write data in an Amazon S3 bucket.

s3ExportConfiguration_bucket :: Lens' S3ExportConfiguration Text

The Amazon S3 bucket name in which a journal export job writes the journal contents.

The bucket name must comply with the Amazon S3 bucket naming conventions. For more information, see Bucket Restrictions and Limitations in the Amazon S3 Developer Guide.

s3ExportConfiguration_prefix :: Lens' S3ExportConfiguration Text

The prefix for the Amazon S3 bucket in which a journal export job writes the journal contents.

The prefix must comply with Amazon S3 key naming rules and restrictions. For more information, see Object Key and Metadata in the Amazon S3 Developer Guide.

The following are examples of valid Prefix values:

JournalExports-ForMyLedger/Testing/

JournalExports

My:Tests/

s3ExportConfiguration_encryptionConfiguration :: Lens' S3ExportConfiguration S3EncryptionConfiguration

The encryption settings that are used by a journal export job to write data in an Amazon S3 bucket.

ValueHolder

data ValueHolder

A structure that can contain a value in multiple encoding formats.

See: newValueHolder smart constructor.

Constructors

ValueHolder'

Fields

ionText :: Maybe (Sensitive Text)

An Amazon Ion plaintext value contained in a ValueHolder structure.

newValueHolder :: ValueHolder

Create a value of ValueHolder with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:ionText:ValueHolder', valueHolder_ionText - An Amazon Ion plaintext value contained in a ValueHolder structure.

valueHolder_ionText :: Lens' ValueHolder (Maybe Text)

An Amazon Ion plaintext value contained in a ValueHolder structure.