Network.AWS.QLDB.Types.JournalKinesisStreamDescription

Description

Documentation

data JournalKinesisStreamDescription

Information about an Amazon QLDB journal stream, including the Amazon Resource Name (ARN), stream name, creation time, current status, and the parameters of the original stream creation request.

See: newJournalKinesisStreamDescription smart constructor.

Constructors

JournalKinesisStreamDescription'

Fields

creationTime :: Maybe POSIX

The date and time, in epoch time format, when the QLDB journal stream was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

arn :: Maybe Text

The Amazon Resource Name (ARN) of the QLDB journal stream.

inclusiveStartTime :: Maybe POSIX

The inclusive start date and time from which to start streaming journal data.

errorCause :: Maybe ErrorCause

The error message that describes the reason that a stream has a status of IMPAIRED or FAILED. This is not applicable to streams that have other status values.

exclusiveEndTime :: Maybe POSIX

The exclusive date and time that specifies when the stream ends. If this parameter is undefined, the stream runs indefinitely until you cancel it.

ledgerName :: Text

The name of the ledger.

roleArn :: Text

The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal stream to write data records to a Kinesis Data Streams resource.

streamId :: Text

The UUID (represented in Base62-encoded text) of the QLDB journal stream.

status :: StreamStatus

The current state of the QLDB journal stream.

kinesisConfiguration :: KinesisConfiguration

The configuration settings of the Amazon Kinesis Data Streams destination for a QLDB journal stream.

streamName :: Text

The user-defined name of the QLDB journal stream.

newJournalKinesisStreamDescription

Arguments

:: Text
$sel:ledgerName:JournalKinesisStreamDescription'

-> Text
$sel:roleArn:JournalKinesisStreamDescription'

-> Text
$sel:streamId:JournalKinesisStreamDescription'

-> StreamStatus
$sel:status:JournalKinesisStreamDescription'

-> KinesisConfiguration
$sel:kinesisConfiguration:JournalKinesisStreamDescription'

-> Text
$sel:streamName:JournalKinesisStreamDescription'

-> JournalKinesisStreamDescription

Create a value of JournalKinesisStreamDescription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:JournalKinesisStreamDescription', journalKinesisStreamDescription_creationTime - The date and time, in epoch time format, when the QLDB journal stream was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

$sel:arn:JournalKinesisStreamDescription', journalKinesisStreamDescription_arn - The Amazon Resource Name (ARN) of the QLDB journal stream.

$sel:inclusiveStartTime:JournalKinesisStreamDescription', journalKinesisStreamDescription_inclusiveStartTime - The inclusive start date and time from which to start streaming journal data.

$sel:errorCause:JournalKinesisStreamDescription', journalKinesisStreamDescription_errorCause - The error message that describes the reason that a stream has a status of IMPAIRED or FAILED. This is not applicable to streams that have other status values.

$sel:exclusiveEndTime:JournalKinesisStreamDescription', journalKinesisStreamDescription_exclusiveEndTime - The exclusive date and time that specifies when the stream ends. If this parameter is undefined, the stream runs indefinitely until you cancel it.

$sel:ledgerName:JournalKinesisStreamDescription', journalKinesisStreamDescription_ledgerName - The name of the ledger.

$sel:roleArn:JournalKinesisStreamDescription', journalKinesisStreamDescription_roleArn - The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal stream to write data records to a Kinesis Data Streams resource.

$sel:streamId:JournalKinesisStreamDescription', journalKinesisStreamDescription_streamId - The UUID (represented in Base62-encoded text) of the QLDB journal stream.

$sel:status:JournalKinesisStreamDescription', journalKinesisStreamDescription_status - The current state of the QLDB journal stream.

$sel:kinesisConfiguration:JournalKinesisStreamDescription', journalKinesisStreamDescription_kinesisConfiguration - The configuration settings of the Amazon Kinesis Data Streams destination for a QLDB journal stream.

$sel:streamName:JournalKinesisStreamDescription', journalKinesisStreamDescription_streamName - The user-defined name of the QLDB journal stream.

journalKinesisStreamDescription_creationTime :: Lens' JournalKinesisStreamDescription (Maybe UTCTime)

The date and time, in epoch time format, when the QLDB journal stream was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

journalKinesisStreamDescription_arn :: Lens' JournalKinesisStreamDescription (Maybe Text)

The Amazon Resource Name (ARN) of the QLDB journal stream.

journalKinesisStreamDescription_inclusiveStartTime :: Lens' JournalKinesisStreamDescription (Maybe UTCTime)

The inclusive start date and time from which to start streaming journal data.

journalKinesisStreamDescription_errorCause :: Lens' JournalKinesisStreamDescription (Maybe ErrorCause)

The error message that describes the reason that a stream has a status of IMPAIRED or FAILED. This is not applicable to streams that have other status values.

journalKinesisStreamDescription_exclusiveEndTime :: Lens' JournalKinesisStreamDescription (Maybe UTCTime)

The exclusive date and time that specifies when the stream ends. If this parameter is undefined, the stream runs indefinitely until you cancel it.

journalKinesisStreamDescription_ledgerName :: Lens' JournalKinesisStreamDescription Text

The name of the ledger.

journalKinesisStreamDescription_roleArn :: Lens' JournalKinesisStreamDescription Text

The Amazon Resource Name (ARN) of the IAM role that grants QLDB permissions for a journal stream to write data records to a Kinesis Data Streams resource.

journalKinesisStreamDescription_streamId :: Lens' JournalKinesisStreamDescription Text

The UUID (represented in Base62-encoded text) of the QLDB journal stream.

journalKinesisStreamDescription_status :: Lens' JournalKinesisStreamDescription StreamStatus

The current state of the QLDB journal stream.

journalKinesisStreamDescription_kinesisConfiguration :: Lens' JournalKinesisStreamDescription KinesisConfiguration

The configuration settings of the Amazon Kinesis Data Streams destination for a QLDB journal stream.

journalKinesisStreamDescription_streamName :: Lens' JournalKinesisStreamDescription Text

The user-defined name of the QLDB journal stream.