Network.AWS.QLDB.UpdateLedger

Description

Updates properties on a ledger.

Creating a Request

data UpdateLedger

See: newUpdateLedger smart constructor.

Constructors

UpdateLedger'

Fields

deletionProtection :: Maybe Bool

The flag that prevents a ledger from being deleted by any user. If not provided on ledger creation, this feature is enabled (true) by default.

If deletion protection is enabled, you must first disable it before you can delete the ledger. You can disable it by calling the UpdateLedger operation to set the flag to false.

kmsKey :: Maybe Text

The key in Key Management Service (KMS) to use for encryption of data at rest in the ledger. For more information, see Encryption at rest in the Amazon QLDB Developer Guide.

Use one of the following options to specify this parameter:

AWS_OWNED_KMS_KEY: Use an KMS key that is owned and managed by Amazon Web Services on your behalf.

Undefined: Make no changes to the KMS key of the ledger.

A valid symmetric customer managed KMS key: Use the specified KMS key in your account that you create, own, and manage.

Amazon QLDB does not support asymmetric keys. For more information, see Using symmetric and asymmetric keys in the Key Management Service Developer Guide.

To specify a customer managed KMS key, you can use its key ID, Amazon Resource Name (ARN), alias name, or alias ARN. When using an alias name, prefix it with "alias/". To specify a key in a different account, you must use the key ARN or alias ARN.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Alias name: alias/ExampleAlias

Alias ARN: arn:aws:kms:us-east-2:111122223333:alias/ExampleAlias

For more information, see Key identifiers (KeyId) in the Key Management Service Developer Guide.

name :: Text

The name of the ledger.

newUpdateLedger

Arguments

:: Text
$sel:name:UpdateLedger'

-> UpdateLedger

Create a value of UpdateLedger with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deletionProtection:UpdateLedger', updateLedger_deletionProtection - The flag that prevents a ledger from being deleted by any user. If not provided on ledger creation, this feature is enabled (true) by default.

If deletion protection is enabled, you must first disable it before you can delete the ledger. You can disable it by calling the UpdateLedger operation to set the flag to false.

$sel:kmsKey:UpdateLedger', updateLedger_kmsKey - The key in Key Management Service (KMS) to use for encryption of data at rest in the ledger. For more information, see Encryption at rest in the Amazon QLDB Developer Guide.

Use one of the following options to specify this parameter:

AWS_OWNED_KMS_KEY: Use an KMS key that is owned and managed by Amazon Web Services on your behalf.

Undefined: Make no changes to the KMS key of the ledger.

A valid symmetric customer managed KMS key: Use the specified KMS key in your account that you create, own, and manage.

Amazon QLDB does not support asymmetric keys. For more information, see Using symmetric and asymmetric keys in the Key Management Service Developer Guide.

To specify a customer managed KMS key, you can use its key ID, Amazon Resource Name (ARN), alias name, or alias ARN. When using an alias name, prefix it with "alias/". To specify a key in a different account, you must use the key ARN or alias ARN.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Alias name: alias/ExampleAlias

Alias ARN: arn:aws:kms:us-east-2:111122223333:alias/ExampleAlias

For more information, see Key identifiers (KeyId) in the Key Management Service Developer Guide.

$sel:name:UpdateLedger', updateLedger_name - The name of the ledger.

Request Lenses

updateLedger_deletionProtection :: Lens' UpdateLedger (Maybe Bool)

The flag that prevents a ledger from being deleted by any user. If not provided on ledger creation, this feature is enabled (true) by default.

If deletion protection is enabled, you must first disable it before you can delete the ledger. You can disable it by calling the UpdateLedger operation to set the flag to false.

updateLedger_kmsKey :: Lens' UpdateLedger (Maybe Text)

The key in Key Management Service (KMS) to use for encryption of data at rest in the ledger. For more information, see Encryption at rest in the Amazon QLDB Developer Guide.

Use one of the following options to specify this parameter:

AWS_OWNED_KMS_KEY: Use an KMS key that is owned and managed by Amazon Web Services on your behalf.

Undefined: Make no changes to the KMS key of the ledger.

A valid symmetric customer managed KMS key: Use the specified KMS key in your account that you create, own, and manage.

Amazon QLDB does not support asymmetric keys. For more information, see Using symmetric and asymmetric keys in the Key Management Service Developer Guide.

To specify a customer managed KMS key, you can use its key ID, Amazon Resource Name (ARN), alias name, or alias ARN. When using an alias name, prefix it with "alias/". To specify a key in a different account, you must use the key ARN or alias ARN.

For example:

Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab

Key ARN: arn:aws:kms:us-east-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab

Alias name: alias/ExampleAlias

Alias ARN: arn:aws:kms:us-east-2:111122223333:alias/ExampleAlias

For more information, see Key identifiers (KeyId) in the Key Management Service Developer Guide.

updateLedger_name :: Lens' UpdateLedger Text

The name of the ledger.

Destructuring the Response

data UpdateLedgerResponse

See: newUpdateLedgerResponse smart constructor.

Constructors

UpdateLedgerResponse'

Fields

state :: Maybe LedgerState

The current status of the ledger.

deletionProtection :: Maybe Bool

The flag that prevents a ledger from being deleted by any user. If not provided on ledger creation, this feature is enabled (true) by default.

If deletion protection is enabled, you must first disable it before you can delete the ledger. You can disable it by calling the UpdateLedger operation to set the flag to false.

arn :: Maybe Text

The Amazon Resource Name (ARN) for the ledger.

encryptionDescription :: Maybe LedgerEncryptionDescription

Information about the encryption of data at rest in the ledger. This includes the current status, the KMS key, and when the key became inaccessible (in the case of an error).

name :: Maybe Text

The name of the ledger.

creationDateTime :: Maybe POSIX

The date and time, in epoch time format, when the ledger was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

httpStatus :: Int

The response's http status code.

newUpdateLedgerResponse

Arguments

:: Int
$sel:httpStatus:UpdateLedgerResponse'

-> UpdateLedgerResponse

Create a value of UpdateLedgerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:state:UpdateLedgerResponse', updateLedgerResponse_state - The current status of the ledger.

$sel:deletionProtection:UpdateLedger', updateLedgerResponse_deletionProtection - The flag that prevents a ledger from being deleted by any user. If not provided on ledger creation, this feature is enabled (true) by default.

If deletion protection is enabled, you must first disable it before you can delete the ledger. You can disable it by calling the UpdateLedger operation to set the flag to false.

$sel:arn:UpdateLedgerResponse', updateLedgerResponse_arn - The Amazon Resource Name (ARN) for the ledger.

$sel:encryptionDescription:UpdateLedgerResponse', updateLedgerResponse_encryptionDescription - Information about the encryption of data at rest in the ledger. This includes the current status, the KMS key, and when the key became inaccessible (in the case of an error).

$sel:name:UpdateLedger', updateLedgerResponse_name - The name of the ledger.

$sel:creationDateTime:UpdateLedgerResponse', updateLedgerResponse_creationDateTime - The date and time, in epoch time format, when the ledger was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

$sel:httpStatus:UpdateLedgerResponse', updateLedgerResponse_httpStatus - The response's http status code.

Response Lenses

updateLedgerResponse_state :: Lens' UpdateLedgerResponse (Maybe LedgerState)

The current status of the ledger.

updateLedgerResponse_deletionProtection :: Lens' UpdateLedgerResponse (Maybe Bool)

The flag that prevents a ledger from being deleted by any user. If not provided on ledger creation, this feature is enabled (true) by default.

If deletion protection is enabled, you must first disable it before you can delete the ledger. You can disable it by calling the UpdateLedger operation to set the flag to false.

updateLedgerResponse_arn :: Lens' UpdateLedgerResponse (Maybe Text)

The Amazon Resource Name (ARN) for the ledger.

updateLedgerResponse_encryptionDescription :: Lens' UpdateLedgerResponse (Maybe LedgerEncryptionDescription)

Information about the encryption of data at rest in the ledger. This includes the current status, the KMS key, and when the key became inaccessible (in the case of an error).

updateLedgerResponse_name :: Lens' UpdateLedgerResponse (Maybe Text)

The name of the ledger.

updateLedgerResponse_creationDateTime :: Lens' UpdateLedgerResponse (Maybe UTCTime)

The date and time, in epoch time format, when the ledger was created. (Epoch time format is the number of seconds elapsed since 12:00:00 AM January 1, 1970 UTC.)

updateLedgerResponse_httpStatus :: Lens' UpdateLedgerResponse Int

The response's http status code.