Network.AWS.QLDB.UpdateLedgerPermissionsMode

Description

Updates the permissions mode of a ledger.

Before you switch to the STANDARD permissions mode, you must first create all required IAM policies and table tags to avoid disruption to your users. To learn more, see Migrating to the standard permissions mode in the Amazon QLDB Developer Guide.

Creating a Request

data UpdateLedgerPermissionsMode

See: newUpdateLedgerPermissionsMode smart constructor.

Constructors

UpdateLedgerPermissionsMode'

Fields

name :: Text

The name of the ledger.

permissionsMode :: PermissionsMode

The permissions mode to assign to the ledger. This parameter can have one of the following values:

ALLOW_ALL: A legacy permissions mode that enables access control with API-level granularity for ledgers.

This mode allows users who have the SendCommand API permission for this ledger to run all PartiQL commands (hence, ALLOW_ALL) on any tables in the specified ledger. This mode disregards any table-level or command-level IAM permissions policies that you create for the ledger.

STANDARD: (Recommended) A permissions mode that enables access control with finer granularity for ledgers, tables, and PartiQL commands.

By default, this mode denies all user requests to run any PartiQL commands on any tables in this ledger. To allow PartiQL commands to run, you must create IAM permissions policies for specific table resources and PartiQL actions, in addition to the SendCommand API permission for the ledger. For information, see Getting started with the standard permissions mode in the Amazon QLDB Developer Guide.

We strongly recommend using the STANDARD permissions mode to maximize the security of your ledger data.

newUpdateLedgerPermissionsMode

Arguments

:: Text
$sel:name:UpdateLedgerPermissionsMode'

-> PermissionsMode
$sel:permissionsMode:UpdateLedgerPermissionsMode'

-> UpdateLedgerPermissionsMode

Create a value of UpdateLedgerPermissionsMode with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:UpdateLedgerPermissionsMode', updateLedgerPermissionsMode_name - The name of the ledger.

$sel:permissionsMode:UpdateLedgerPermissionsMode', updateLedgerPermissionsMode_permissionsMode - The permissions mode to assign to the ledger. This parameter can have one of the following values:

ALLOW_ALL: A legacy permissions mode that enables access control with API-level granularity for ledgers.

This mode allows users who have the SendCommand API permission for this ledger to run all PartiQL commands (hence, ALLOW_ALL) on any tables in the specified ledger. This mode disregards any table-level or command-level IAM permissions policies that you create for the ledger.

STANDARD: (Recommended) A permissions mode that enables access control with finer granularity for ledgers, tables, and PartiQL commands.

By default, this mode denies all user requests to run any PartiQL commands on any tables in this ledger. To allow PartiQL commands to run, you must create IAM permissions policies for specific table resources and PartiQL actions, in addition to the SendCommand API permission for the ledger. For information, see Getting started with the standard permissions mode in the Amazon QLDB Developer Guide.

We strongly recommend using the STANDARD permissions mode to maximize the security of your ledger data.

Request Lenses

updateLedgerPermissionsMode_name :: Lens' UpdateLedgerPermissionsMode Text

The name of the ledger.

updateLedgerPermissionsMode_permissionsMode :: Lens' UpdateLedgerPermissionsMode PermissionsMode

The permissions mode to assign to the ledger. This parameter can have one of the following values:

ALLOW_ALL: A legacy permissions mode that enables access control with API-level granularity for ledgers.

This mode allows users who have the SendCommand API permission for this ledger to run all PartiQL commands (hence, ALLOW_ALL) on any tables in the specified ledger. This mode disregards any table-level or command-level IAM permissions policies that you create for the ledger.

STANDARD: (Recommended) A permissions mode that enables access control with finer granularity for ledgers, tables, and PartiQL commands.

By default, this mode denies all user requests to run any PartiQL commands on any tables in this ledger. To allow PartiQL commands to run, you must create IAM permissions policies for specific table resources and PartiQL actions, in addition to the SendCommand API permission for the ledger. For information, see Getting started with the standard permissions mode in the Amazon QLDB Developer Guide.

We strongly recommend using the STANDARD permissions mode to maximize the security of your ledger data.

Destructuring the Response

data UpdateLedgerPermissionsModeResponse

See: newUpdateLedgerPermissionsModeResponse smart constructor.

Constructors

UpdateLedgerPermissionsModeResponse'

Fields

permissionsMode :: Maybe PermissionsMode

The current permissions mode of the ledger.

arn :: Maybe Text

The Amazon Resource Name (ARN) for the ledger.

name :: Maybe Text

The name of the ledger.

httpStatus :: Int

The response's http status code.

newUpdateLedgerPermissionsModeResponse

Arguments

:: Int
$sel:httpStatus:UpdateLedgerPermissionsModeResponse'

-> UpdateLedgerPermissionsModeResponse

Create a value of UpdateLedgerPermissionsModeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:permissionsMode:UpdateLedgerPermissionsMode', updateLedgerPermissionsModeResponse_permissionsMode - The current permissions mode of the ledger.

$sel:arn:UpdateLedgerPermissionsModeResponse', updateLedgerPermissionsModeResponse_arn - The Amazon Resource Name (ARN) for the ledger.

$sel:name:UpdateLedgerPermissionsMode', updateLedgerPermissionsModeResponse_name - The name of the ledger.

$sel:httpStatus:UpdateLedgerPermissionsModeResponse', updateLedgerPermissionsModeResponse_httpStatus - The response's http status code.

Response Lenses

updateLedgerPermissionsModeResponse_permissionsMode :: Lens' UpdateLedgerPermissionsModeResponse (Maybe PermissionsMode)

The current permissions mode of the ledger.

updateLedgerPermissionsModeResponse_arn :: Lens' UpdateLedgerPermissionsModeResponse (Maybe Text)

The Amazon Resource Name (ARN) for the ledger.

updateLedgerPermissionsModeResponse_name :: Lens' UpdateLedgerPermissionsModeResponse (Maybe Text)

The name of the ledger.

updateLedgerPermissionsModeResponse_httpStatus :: Lens' UpdateLedgerPermissionsModeResponse Int

The response's http status code.