Network.AWS.QLDBSession.Lens

Description

Operations

SendCommand

sendCommand_fetchPage :: Lens' SendCommand (Maybe FetchPageRequest)

Command to fetch a page.

sendCommand_sessionToken :: Lens' SendCommand (Maybe Text)

Specifies the session token for the current command. A session token is constant throughout the life of the session.

To obtain a session token, run the StartSession command. This SessionToken is required for every subsequent command that is issued during the current session.

sendCommand_abortTransaction :: Lens' SendCommand (Maybe AbortTransactionRequest)

Command to abort the current transaction.

sendCommand_endSession :: Lens' SendCommand (Maybe EndSessionRequest)

Command to end the current session.

sendCommand_startTransaction :: Lens' SendCommand (Maybe StartTransactionRequest)

Command to start a new transaction.

sendCommand_executeStatement :: Lens' SendCommand (Maybe ExecuteStatementRequest)

Command to execute a statement in the specified transaction.

sendCommand_commitTransaction :: Lens' SendCommand (Maybe CommitTransactionRequest)

Command to commit the specified transaction.

sendCommand_startSession :: Lens' SendCommand (Maybe StartSessionRequest)

Command to start a new session. A session token is obtained as part of the response.

sendCommandResponse_fetchPage :: Lens' SendCommandResponse (Maybe FetchPageResult)

Contains the details of the fetched page.

sendCommandResponse_abortTransaction :: Lens' SendCommandResponse (Maybe AbortTransactionResult)

Contains the details of the aborted transaction.

sendCommandResponse_endSession :: Lens' SendCommandResponse (Maybe EndSessionResult)

Contains the details of the ended session.

sendCommandResponse_startTransaction :: Lens' SendCommandResponse (Maybe StartTransactionResult)

Contains the details of the started transaction.

sendCommandResponse_executeStatement :: Lens' SendCommandResponse (Maybe ExecuteStatementResult)

Contains the details of the executed statement.

sendCommandResponse_commitTransaction :: Lens' SendCommandResponse (Maybe CommitTransactionResult)

Contains the details of the committed transaction.

sendCommandResponse_startSession :: Lens' SendCommandResponse (Maybe StartSessionResult)

Contains the details of the started session that includes a session token. This SessionToken is required for every subsequent command that is issued during the current session.

sendCommandResponse_httpStatus :: Lens' SendCommandResponse Int

The response's http status code.

Types

AbortTransactionRequest

AbortTransactionResult

abortTransactionResult_timingInformation :: Lens' AbortTransactionResult (Maybe TimingInformation)

Contains server-side performance information for the command.

CommitTransactionRequest

commitTransactionRequest_transactionId :: Lens' CommitTransactionRequest Text

Specifies the transaction ID of the transaction to commit.

commitTransactionRequest_commitDigest :: Lens' CommitTransactionRequest ByteString

Specifies the commit digest for the transaction to commit. For every active transaction, the commit digest must be passed. QLDB validates CommitDigest and rejects the commit with an error if the digest computed on the client does not match the digest computed by QLDB.

The purpose of the CommitDigest parameter is to ensure that QLDB commits a transaction if and only if the server has processed the exact set of statements sent by the client, in the same order that client sent them, and with no duplicates.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

CommitTransactionResult

commitTransactionResult_timingInformation :: Lens' CommitTransactionResult (Maybe TimingInformation)

Contains server-side performance information for the command.

commitTransactionResult_consumedIOs :: Lens' CommitTransactionResult (Maybe IOUsage)

Contains metrics about the number of I/O requests that were consumed.

commitTransactionResult_commitDigest :: Lens' CommitTransactionResult (Maybe ByteString)

The commit digest of the committed transaction.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

commitTransactionResult_transactionId :: Lens' CommitTransactionResult (Maybe Text)

The transaction ID of the committed transaction.

EndSessionRequest

EndSessionResult

endSessionResult_timingInformation :: Lens' EndSessionResult (Maybe TimingInformation)

Contains server-side performance information for the command.

ExecuteStatementRequest

executeStatementRequest_parameters :: Lens' ExecuteStatementRequest (Maybe [ValueHolder])

Specifies the parameters for the parameterized statement in the request.

executeStatementRequest_transactionId :: Lens' ExecuteStatementRequest Text

Specifies the transaction ID of the request.

executeStatementRequest_statement :: Lens' ExecuteStatementRequest Text

Specifies the statement of the request.

ExecuteStatementResult

executeStatementResult_timingInformation :: Lens' ExecuteStatementResult (Maybe TimingInformation)

Contains server-side performance information for the command.

executeStatementResult_consumedIOs :: Lens' ExecuteStatementResult (Maybe IOUsage)

Contains metrics about the number of I/O requests that were consumed.

executeStatementResult_firstPage :: Lens' ExecuteStatementResult (Maybe Page)

Contains the details of the first fetched page.

FetchPageRequest

fetchPageRequest_transactionId :: Lens' FetchPageRequest Text

Specifies the transaction ID of the page to be fetched.

fetchPageRequest_nextPageToken :: Lens' FetchPageRequest Text

Specifies the next page token of the page to be fetched.

FetchPageResult

fetchPageResult_timingInformation :: Lens' FetchPageResult (Maybe TimingInformation)

Contains server-side performance information for the command.

fetchPageResult_consumedIOs :: Lens' FetchPageResult (Maybe IOUsage)

Contains metrics about the number of I/O requests that were consumed.

fetchPageResult_page :: Lens' FetchPageResult (Maybe Page)

Contains details of the fetched page.

IOUsage

iOUsage_readIOs :: Lens' IOUsage (Maybe Integer)

The number of read I/O requests that the command made.

iOUsage_writeIOs :: Lens' IOUsage (Maybe Integer)

The number of write I/O requests that the command made.

Page

page_nextPageToken :: Lens' Page (Maybe Text)

The token of the next page.

page_values :: Lens' Page (Maybe [ValueHolder])

A structure that contains values in multiple encoding formats.

StartSessionRequest

startSessionRequest_ledgerName :: Lens' StartSessionRequest Text

The name of the ledger to start a new session against.

StartSessionResult

startSessionResult_timingInformation :: Lens' StartSessionResult (Maybe TimingInformation)

Contains server-side performance information for the command.

startSessionResult_sessionToken :: Lens' StartSessionResult (Maybe Text)

Session token of the started session. This SessionToken is required for every subsequent command that is issued during the current session.

StartTransactionRequest

StartTransactionResult

startTransactionResult_timingInformation :: Lens' StartTransactionResult (Maybe TimingInformation)

Contains server-side performance information for the command.

startTransactionResult_transactionId :: Lens' StartTransactionResult (Maybe Text)

The transaction ID of the started transaction.

TimingInformation

timingInformation_processingTimeMilliseconds :: Lens' TimingInformation (Maybe Integer)

The amount of time that QLDB spent on processing the command, measured in milliseconds.

ValueHolder

valueHolder_ionText :: Lens' ValueHolder (Maybe Text)

An Amazon Ion plaintext value contained in a ValueHolder structure.

valueHolder_ionBinary :: Lens' ValueHolder (Maybe ByteString)

An Amazon Ion binary value contained in a ValueHolder structure.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.