Network.AWS.QLDBSession.Types.CommitTransactionResult

Description

Documentation

data CommitTransactionResult

Contains the details of the committed transaction.

See: newCommitTransactionResult smart constructor.

Constructors

CommitTransactionResult'

Fields

timingInformation :: Maybe TimingInformation

Contains server-side performance information for the command.

consumedIOs :: Maybe IOUsage

Contains metrics about the number of I/O requests that were consumed.

commitDigest :: Maybe Base64

The commit digest of the committed transaction.

transactionId :: Maybe Text

The transaction ID of the committed transaction.

newCommitTransactionResult :: CommitTransactionResult

Create a value of CommitTransactionResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:timingInformation:CommitTransactionResult', commitTransactionResult_timingInformation - Contains server-side performance information for the command.

$sel:consumedIOs:CommitTransactionResult', commitTransactionResult_consumedIOs - Contains metrics about the number of I/O requests that were consumed.

$sel:commitDigest:CommitTransactionResult', commitTransactionResult_commitDigest - The commit digest of the committed transaction.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:transactionId:CommitTransactionResult', commitTransactionResult_transactionId - The transaction ID of the committed transaction.

commitTransactionResult_timingInformation :: Lens' CommitTransactionResult (Maybe TimingInformation)

Contains server-side performance information for the command.

commitTransactionResult_consumedIOs :: Lens' CommitTransactionResult (Maybe IOUsage)

Contains metrics about the number of I/O requests that were consumed.

commitTransactionResult_commitDigest :: Lens' CommitTransactionResult (Maybe ByteString)

The commit digest of the committed transaction.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

commitTransactionResult_transactionId :: Lens' CommitTransactionResult (Maybe Text)

The transaction ID of the committed transaction.