Network.AWS.QLDBSession.Types.ExecuteStatementRequest

Description

Documentation

data ExecuteStatementRequest

Specifies a request to execute a statement.

See: newExecuteStatementRequest smart constructor.

Constructors

ExecuteStatementRequest'

Fields

parameters :: Maybe [ValueHolder]

Specifies the parameters for the parameterized statement in the request.

transactionId :: Text

Specifies the transaction ID of the request.

statement :: Text

Specifies the statement of the request.

newExecuteStatementRequest

Arguments

:: Text
$sel:transactionId:ExecuteStatementRequest'

-> Text
$sel:statement:ExecuteStatementRequest'

-> ExecuteStatementRequest

Create a value of ExecuteStatementRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameters:ExecuteStatementRequest', executeStatementRequest_parameters - Specifies the parameters for the parameterized statement in the request.

$sel:transactionId:ExecuteStatementRequest', executeStatementRequest_transactionId - Specifies the transaction ID of the request.

$sel:statement:ExecuteStatementRequest', executeStatementRequest_statement - Specifies the statement of the request.

executeStatementRequest_parameters :: Lens' ExecuteStatementRequest (Maybe [ValueHolder])

Specifies the parameters for the parameterized statement in the request.

executeStatementRequest_transactionId :: Lens' ExecuteStatementRequest Text

Specifies the transaction ID of the request.

executeStatementRequest_statement :: Lens' ExecuteStatementRequest Text

Specifies the statement of the request.