Network.AWS.QLDBSession.Types.FetchPageRequest

Description

Documentation

data FetchPageRequest

Specifies the details of the page to be fetched.

See: newFetchPageRequest smart constructor.

Constructors

FetchPageRequest'

Fields

transactionId :: Text

Specifies the transaction ID of the page to be fetched.

nextPageToken :: Text

Specifies the next page token of the page to be fetched.

newFetchPageRequest

Arguments

:: Text
$sel:transactionId:FetchPageRequest'

-> Text
$sel:nextPageToken:FetchPageRequest'

-> FetchPageRequest

Create a value of FetchPageRequest with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transactionId:FetchPageRequest', fetchPageRequest_transactionId - Specifies the transaction ID of the page to be fetched.

$sel:nextPageToken:FetchPageRequest', fetchPageRequest_nextPageToken - Specifies the next page token of the page to be fetched.

fetchPageRequest_transactionId :: Lens' FetchPageRequest Text

Specifies the transaction ID of the page to be fetched.

fetchPageRequest_nextPageToken :: Lens' FetchPageRequest Text

Specifies the next page token of the page to be fetched.