Network.AWS.QuickSight.CreateAccountCustomization

Description

Creates Amazon QuickSight customizations the current Amazon Web Services Region. Currently, you can add a custom default theme by using the CreateAccountCustomization or UpdateAccountCustomization API operation. To further customize Amazon QuickSight by removing Amazon QuickSight sample assets and videos for all new users, see Customizing Amazon QuickSight in the Amazon QuickSight User Guide.

You can create customizations for your Amazon Web Services account or, if you specify a namespace, for a Amazon QuickSight namespace instead. Customizations that apply to a namespace always override customizations that apply to an Amazon Web Services account. To find out which customizations apply, use the DescribeAccountCustomization API operation.

Before you use the CreateAccountCustomization API operation to add a theme as the namespace default, make sure that you first share the theme with the namespace. If you don't share it with the namespace, the theme isn't visible to your users even if you make it the default theme. To check if the theme is shared, view the current permissions by using the DescribeThemePermissions API operation. To share the theme, grant permissions by using the UpdateThemePermissions API operation.

Creating a Request

data CreateAccountCustomization

See: newCreateAccountCustomization smart constructor.

Constructors

CreateAccountCustomization'

Fields

namespace :: Maybe Text

The Amazon QuickSight namespace that you want to add customizations to.

tags :: Maybe (NonEmpty Tag)

A list of the tags that you want to attach to this resource.

awsAccountId :: Text

The ID for the Amazon Web Services account that you want to customize Amazon QuickSight for.

accountCustomization :: AccountCustomization

The Amazon QuickSight customizations you're adding in the current Amazon Web Services Region. You can add these to an Amazon Web Services account and a Amazon QuickSight namespace.

For example, you can add a default theme by setting AccountCustomization to the midnight theme: "AccountCustomization": { "DefaultTheme": "arn:aws:quicksight::aws:theme/MIDNIGHT" }. Or, you can add a custom theme by specifying "AccountCustomization": { "DefaultTheme": "arn:aws:quicksight:us-west-2:111122223333:theme/bdb844d0-0fe9-4d9d-b520-0fe602d93639" }.

newCreateAccountCustomization

Arguments

:: Text
$sel:awsAccountId:CreateAccountCustomization'

-> AccountCustomization
$sel:accountCustomization:CreateAccountCustomization'

-> CreateAccountCustomization

Create a value of CreateAccountCustomization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:namespace:CreateAccountCustomization', createAccountCustomization_namespace - The Amazon QuickSight namespace that you want to add customizations to.

$sel:tags:CreateAccountCustomization', createAccountCustomization_tags - A list of the tags that you want to attach to this resource.

$sel:awsAccountId:CreateAccountCustomization', createAccountCustomization_awsAccountId - The ID for the Amazon Web Services account that you want to customize Amazon QuickSight for.

$sel:accountCustomization:CreateAccountCustomization', createAccountCustomization_accountCustomization - The Amazon QuickSight customizations you're adding in the current Amazon Web Services Region. You can add these to an Amazon Web Services account and a Amazon QuickSight namespace.

For example, you can add a default theme by setting AccountCustomization to the midnight theme: "AccountCustomization": { "DefaultTheme": "arn:aws:quicksight::aws:theme/MIDNIGHT" }. Or, you can add a custom theme by specifying "AccountCustomization": { "DefaultTheme": "arn:aws:quicksight:us-west-2:111122223333:theme/bdb844d0-0fe9-4d9d-b520-0fe602d93639" }.

Request Lenses

createAccountCustomization_namespace :: Lens' CreateAccountCustomization (Maybe Text)

The Amazon QuickSight namespace that you want to add customizations to.

createAccountCustomization_tags :: Lens' CreateAccountCustomization (Maybe (NonEmpty Tag))

A list of the tags that you want to attach to this resource.

createAccountCustomization_awsAccountId :: Lens' CreateAccountCustomization Text

The ID for the Amazon Web Services account that you want to customize Amazon QuickSight for.

createAccountCustomization_accountCustomization :: Lens' CreateAccountCustomization AccountCustomization

The Amazon QuickSight customizations you're adding in the current Amazon Web Services Region. You can add these to an Amazon Web Services account and a Amazon QuickSight namespace.

For example, you can add a default theme by setting AccountCustomization to the midnight theme: "AccountCustomization": { "DefaultTheme": "arn:aws:quicksight::aws:theme/MIDNIGHT" }. Or, you can add a custom theme by specifying "AccountCustomization": { "DefaultTheme": "arn:aws:quicksight:us-west-2:111122223333:theme/bdb844d0-0fe9-4d9d-b520-0fe602d93639" }.

Destructuring the Response

data CreateAccountCustomizationResponse

See: newCreateAccountCustomizationResponse smart constructor.

Constructors

CreateAccountCustomizationResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

accountCustomization :: Maybe AccountCustomization

The Amazon QuickSight customizations you're adding in the current Amazon Web Services Region.

arn :: Maybe Text

The Amazon Resource Name (ARN) for the customization that you created for this Amazon Web Services account.

namespace :: Maybe Text

The namespace associated with the customization you're creating.

awsAccountId :: Maybe Text

The ID for the Amazon Web Services account that you want to customize Amazon QuickSight for.

status :: Int

The HTTP status of the request.

newCreateAccountCustomizationResponse

Arguments

:: Int
$sel:status:CreateAccountCustomizationResponse'

-> CreateAccountCustomizationResponse

Create a value of CreateAccountCustomizationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:CreateAccountCustomizationResponse', createAccountCustomizationResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:accountCustomization:CreateAccountCustomization', createAccountCustomizationResponse_accountCustomization - The Amazon QuickSight customizations you're adding in the current Amazon Web Services Region.

$sel:arn:CreateAccountCustomizationResponse', createAccountCustomizationResponse_arn - The Amazon Resource Name (ARN) for the customization that you created for this Amazon Web Services account.

$sel:namespace:CreateAccountCustomization', createAccountCustomizationResponse_namespace - The namespace associated with the customization you're creating.

$sel:awsAccountId:CreateAccountCustomization', createAccountCustomizationResponse_awsAccountId - The ID for the Amazon Web Services account that you want to customize Amazon QuickSight for.

$sel:status:CreateAccountCustomizationResponse', createAccountCustomizationResponse_status - The HTTP status of the request.

Response Lenses

createAccountCustomizationResponse_requestId :: Lens' CreateAccountCustomizationResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

createAccountCustomizationResponse_accountCustomization :: Lens' CreateAccountCustomizationResponse (Maybe AccountCustomization)

The Amazon QuickSight customizations you're adding in the current Amazon Web Services Region.

createAccountCustomizationResponse_arn :: Lens' CreateAccountCustomizationResponse (Maybe Text)

The Amazon Resource Name (ARN) for the customization that you created for this Amazon Web Services account.

createAccountCustomizationResponse_namespace :: Lens' CreateAccountCustomizationResponse (Maybe Text)

The namespace associated with the customization you're creating.

createAccountCustomizationResponse_awsAccountId :: Lens' CreateAccountCustomizationResponse (Maybe Text)

The ID for the Amazon Web Services account that you want to customize Amazon QuickSight for.

createAccountCustomizationResponse_status :: Lens' CreateAccountCustomizationResponse Int

The HTTP status of the request.