Network.AWS.QuickSight.CreateAnalysis

Description

Creates an analysis in Amazon QuickSight.

Creating a Request

data CreateAnalysis

See: newCreateAnalysis smart constructor.

Constructors

CreateAnalysis'

Fields

themeArn :: Maybe Text

The ARN for the theme to apply to the analysis that you're creating. To see the theme in the Amazon QuickSight console, make sure that you have access to it.

parameters :: Maybe Parameters

The parameter names and override values that you want to use. An analysis can have any parameter type, and some parameters might accept multiple values.

permissions :: Maybe (NonEmpty ResourcePermission)

A structure that describes the principals and the resource-level permissions on an analysis. You can use the Permissions structure to grant permissions by providing a list of Identity and Access Management (IAM) action information for each principal listed by Amazon Resource Name (ARN).

To specify no permissions, omit Permissions.

tags :: Maybe (NonEmpty Tag)

Contains a map of the key-value pairs for the resource tag or tags assigned to the analysis.

awsAccountId :: Text

The ID of the Amazon Web Services account where you are creating an analysis.

analysisId :: Text

The ID for the analysis that you're creating. This ID displays in the URL of the analysis.

name :: Text

A descriptive name for the analysis that you're creating. This name displays for the analysis in the Amazon QuickSight console.

sourceEntity :: AnalysisSourceEntity

A source entity to use for the analysis that you're creating. This metadata structure contains details that describe a source template and one or more datasets.

newCreateAnalysis

Arguments

:: Text
$sel:awsAccountId:CreateAnalysis'

-> Text
$sel:analysisId:CreateAnalysis'

-> Text
$sel:name:CreateAnalysis'

-> AnalysisSourceEntity
$sel:sourceEntity:CreateAnalysis'

-> CreateAnalysis

Create a value of CreateAnalysis with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:themeArn:CreateAnalysis', createAnalysis_themeArn - The ARN for the theme to apply to the analysis that you're creating. To see the theme in the Amazon QuickSight console, make sure that you have access to it.

$sel:parameters:CreateAnalysis', createAnalysis_parameters - The parameter names and override values that you want to use. An analysis can have any parameter type, and some parameters might accept multiple values.

$sel:permissions:CreateAnalysis', createAnalysis_permissions - A structure that describes the principals and the resource-level permissions on an analysis. You can use the Permissions structure to grant permissions by providing a list of Identity and Access Management (IAM) action information for each principal listed by Amazon Resource Name (ARN).

To specify no permissions, omit Permissions.

$sel:tags:CreateAnalysis', createAnalysis_tags - Contains a map of the key-value pairs for the resource tag or tags assigned to the analysis.

$sel:awsAccountId:CreateAnalysis', createAnalysis_awsAccountId - The ID of the Amazon Web Services account where you are creating an analysis.

$sel:analysisId:CreateAnalysis', createAnalysis_analysisId - The ID for the analysis that you're creating. This ID displays in the URL of the analysis.

$sel:name:CreateAnalysis', createAnalysis_name - A descriptive name for the analysis that you're creating. This name displays for the analysis in the Amazon QuickSight console.

$sel:sourceEntity:CreateAnalysis', createAnalysis_sourceEntity - A source entity to use for the analysis that you're creating. This metadata structure contains details that describe a source template and one or more datasets.

Request Lenses

createAnalysis_themeArn :: Lens' CreateAnalysis (Maybe Text)

The ARN for the theme to apply to the analysis that you're creating. To see the theme in the Amazon QuickSight console, make sure that you have access to it.

createAnalysis_parameters :: Lens' CreateAnalysis (Maybe Parameters)

The parameter names and override values that you want to use. An analysis can have any parameter type, and some parameters might accept multiple values.

createAnalysis_permissions :: Lens' CreateAnalysis (Maybe (NonEmpty ResourcePermission))

A structure that describes the principals and the resource-level permissions on an analysis. You can use the Permissions structure to grant permissions by providing a list of Identity and Access Management (IAM) action information for each principal listed by Amazon Resource Name (ARN).

To specify no permissions, omit Permissions.

createAnalysis_tags :: Lens' CreateAnalysis (Maybe (NonEmpty Tag))

Contains a map of the key-value pairs for the resource tag or tags assigned to the analysis.

createAnalysis_awsAccountId :: Lens' CreateAnalysis Text

The ID of the Amazon Web Services account where you are creating an analysis.

createAnalysis_analysisId :: Lens' CreateAnalysis Text

The ID for the analysis that you're creating. This ID displays in the URL of the analysis.

createAnalysis_name :: Lens' CreateAnalysis Text

A descriptive name for the analysis that you're creating. This name displays for the analysis in the Amazon QuickSight console.

createAnalysis_sourceEntity :: Lens' CreateAnalysis AnalysisSourceEntity

A source entity to use for the analysis that you're creating. This metadata structure contains details that describe a source template and one or more datasets.

Destructuring the Response

data CreateAnalysisResponse

See: newCreateAnalysisResponse smart constructor.

Constructors

CreateAnalysisResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

analysisId :: Maybe Text

The ID of the analysis.

arn :: Maybe Text

The ARN for the analysis.

creationStatus :: Maybe ResourceStatus

The status of the creation of the analysis.

status :: Int

The HTTP status of the request.

newCreateAnalysisResponse

Arguments

:: Int
$sel:status:CreateAnalysisResponse'

-> CreateAnalysisResponse

Create a value of CreateAnalysisResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:CreateAnalysisResponse', createAnalysisResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:analysisId:CreateAnalysis', createAnalysisResponse_analysisId - The ID of the analysis.

$sel:arn:CreateAnalysisResponse', createAnalysisResponse_arn - The ARN for the analysis.

$sel:creationStatus:CreateAnalysisResponse', createAnalysisResponse_creationStatus - The status of the creation of the analysis.

$sel:status:CreateAnalysisResponse', createAnalysisResponse_status - The HTTP status of the request.

Response Lenses

createAnalysisResponse_requestId :: Lens' CreateAnalysisResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

createAnalysisResponse_analysisId :: Lens' CreateAnalysisResponse (Maybe Text)

The ID of the analysis.

createAnalysisResponse_arn :: Lens' CreateAnalysisResponse (Maybe Text)

The ARN for the analysis.

createAnalysisResponse_creationStatus :: Lens' CreateAnalysisResponse (Maybe ResourceStatus)

The status of the creation of the analysis.

createAnalysisResponse_status :: Lens' CreateAnalysisResponse Int

The HTTP status of the request.