Network.AWS.QuickSight.CreateDashboard

Description

Creates a dashboard from a template. To first create a template, see the CreateTemplate API operation.

A dashboard is an entity in Amazon QuickSight that identifies Amazon QuickSight reports, created from analyses. You can share Amazon QuickSight dashboards. With the right permissions, you can create scheduled email reports from them. If you have the correct permissions, you can create a dashboard from a template that exists in a different Amazon Web Services account.

Creating a Request

data CreateDashboard

See: newCreateDashboard smart constructor.

Constructors

CreateDashboard'

Fields

themeArn :: Maybe Text

The Amazon Resource Name (ARN) of the theme that is being used for this dashboard. If you add a value for this field, it overrides the value that is used in the source entity. The theme ARN must exist in the same Amazon Web Services account where you create the dashboard.

dashboardPublishOptions :: Maybe DashboardPublishOptions

Options for publishing the dashboard when you create it:

AvailabilityStatus for AdHocFilteringOption - This status can be either ENABLED or DISABLED. When this is set to DISABLED, Amazon QuickSight disables the left filter pane on the published dashboard, which can be used for ad hoc (one-time) filtering. This option is ENABLED by default.

AvailabilityStatus for ExportToCSVOption - This status can be either ENABLED or DISABLED. The visual option to export data to .CSV format isn't enabled when this is set to DISABLED. This option is ENABLED by default.

VisibilityState for SheetControlsOption - This visibility state can be either COLLAPSED or EXPANDED. This option is COLLAPSED by default.

versionDescription :: Maybe Text

A description for the first version of the dashboard being created.

parameters :: Maybe Parameters

The parameters for the creation of the dashboard, which you want to use to override the default settings. A dashboard can have any type of parameters, and some parameters might accept multiple values.

permissions :: Maybe (NonEmpty ResourcePermission)

A structure that contains the permissions of the dashboard. You can use this structure for granting permissions by providing a list of IAMaction information for each principal ARN.

To specify no permissions, omit the permissions list.

tags :: Maybe (NonEmpty Tag)

Contains a map of the key-value pairs for the resource tag or tags assigned to the dashboard.

awsAccountId :: Text

The ID of the Amazon Web Services account where you want to create the dashboard.

dashboardId :: Text

The ID for the dashboard, also added to the IAMpolicy.

name :: Text

The display name of the dashboard.

sourceEntity :: DashboardSourceEntity

The entity that you are using as a source when you create the dashboard. In SourceEntity, you specify the type of object you're using as source. You can only create a dashboard from a template, so you use a SourceTemplate entity. If you need to create a dashboard from an analysis, first convert the analysis to a template by using the CreateTemplate API operation. For SourceTemplate, specify the Amazon Resource Name (ARN) of the source template. The SourceTemplateARN can contain any Amazon Web Services account and any Amazon QuickSight-supported Amazon Web Services Region.

Use the DataSetReferences entity within SourceTemplate to list the replacement datasets for the placeholders listed in the original. The schema in each dataset must match its placeholder.

newCreateDashboard

Arguments

:: Text
$sel:awsAccountId:CreateDashboard'

-> Text
$sel:dashboardId:CreateDashboard'

-> Text
$sel:name:CreateDashboard'

-> DashboardSourceEntity
$sel:sourceEntity:CreateDashboard'

-> CreateDashboard

Create a value of CreateDashboard with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:themeArn:CreateDashboard', createDashboard_themeArn - The Amazon Resource Name (ARN) of the theme that is being used for this dashboard. If you add a value for this field, it overrides the value that is used in the source entity. The theme ARN must exist in the same Amazon Web Services account where you create the dashboard.

$sel:dashboardPublishOptions:CreateDashboard', createDashboard_dashboardPublishOptions - Options for publishing the dashboard when you create it:

AvailabilityStatus for AdHocFilteringOption - This status can be either ENABLED or DISABLED. When this is set to DISABLED, Amazon QuickSight disables the left filter pane on the published dashboard, which can be used for ad hoc (one-time) filtering. This option is ENABLED by default.

AvailabilityStatus for ExportToCSVOption - This status can be either ENABLED or DISABLED. The visual option to export data to .CSV format isn't enabled when this is set to DISABLED. This option is ENABLED by default.

VisibilityState for SheetControlsOption - This visibility state can be either COLLAPSED or EXPANDED. This option is COLLAPSED by default.

$sel:versionDescription:CreateDashboard', createDashboard_versionDescription - A description for the first version of the dashboard being created.

$sel:parameters:CreateDashboard', createDashboard_parameters - The parameters for the creation of the dashboard, which you want to use to override the default settings. A dashboard can have any type of parameters, and some parameters might accept multiple values.

$sel:permissions:CreateDashboard', createDashboard_permissions - A structure that contains the permissions of the dashboard. You can use this structure for granting permissions by providing a list of IAMaction information for each principal ARN.

To specify no permissions, omit the permissions list.

$sel:tags:CreateDashboard', createDashboard_tags - Contains a map of the key-value pairs for the resource tag or tags assigned to the dashboard.

$sel:awsAccountId:CreateDashboard', createDashboard_awsAccountId - The ID of the Amazon Web Services account where you want to create the dashboard.

$sel:dashboardId:CreateDashboard', createDashboard_dashboardId - The ID for the dashboard, also added to the IAMpolicy.

$sel:name:CreateDashboard', createDashboard_name - The display name of the dashboard.

$sel:sourceEntity:CreateDashboard', createDashboard_sourceEntity - The entity that you are using as a source when you create the dashboard. In SourceEntity, you specify the type of object you're using as source. You can only create a dashboard from a template, so you use a SourceTemplate entity. If you need to create a dashboard from an analysis, first convert the analysis to a template by using the CreateTemplate API operation. For SourceTemplate, specify the Amazon Resource Name (ARN) of the source template. The SourceTemplateARN can contain any Amazon Web Services account and any Amazon QuickSight-supported Amazon Web Services Region.

Use the DataSetReferences entity within SourceTemplate to list the replacement datasets for the placeholders listed in the original. The schema in each dataset must match its placeholder.

Request Lenses

createDashboard_themeArn :: Lens' CreateDashboard (Maybe Text)

The Amazon Resource Name (ARN) of the theme that is being used for this dashboard. If you add a value for this field, it overrides the value that is used in the source entity. The theme ARN must exist in the same Amazon Web Services account where you create the dashboard.

createDashboard_dashboardPublishOptions :: Lens' CreateDashboard (Maybe DashboardPublishOptions)

Options for publishing the dashboard when you create it:

AvailabilityStatus for AdHocFilteringOption - This status can be either ENABLED or DISABLED. When this is set to DISABLED, Amazon QuickSight disables the left filter pane on the published dashboard, which can be used for ad hoc (one-time) filtering. This option is ENABLED by default.

AvailabilityStatus for ExportToCSVOption - This status can be either ENABLED or DISABLED. The visual option to export data to .CSV format isn't enabled when this is set to DISABLED. This option is ENABLED by default.

VisibilityState for SheetControlsOption - This visibility state can be either COLLAPSED or EXPANDED. This option is COLLAPSED by default.

createDashboard_versionDescription :: Lens' CreateDashboard (Maybe Text)

A description for the first version of the dashboard being created.

createDashboard_parameters :: Lens' CreateDashboard (Maybe Parameters)

The parameters for the creation of the dashboard, which you want to use to override the default settings. A dashboard can have any type of parameters, and some parameters might accept multiple values.

createDashboard_permissions :: Lens' CreateDashboard (Maybe (NonEmpty ResourcePermission))

A structure that contains the permissions of the dashboard. You can use this structure for granting permissions by providing a list of IAMaction information for each principal ARN.

To specify no permissions, omit the permissions list.

createDashboard_tags :: Lens' CreateDashboard (Maybe (NonEmpty Tag))

Contains a map of the key-value pairs for the resource tag or tags assigned to the dashboard.

createDashboard_awsAccountId :: Lens' CreateDashboard Text

The ID of the Amazon Web Services account where you want to create the dashboard.

createDashboard_dashboardId :: Lens' CreateDashboard Text

The ID for the dashboard, also added to the IAMpolicy.

createDashboard_name :: Lens' CreateDashboard Text

The display name of the dashboard.

createDashboard_sourceEntity :: Lens' CreateDashboard DashboardSourceEntity

The entity that you are using as a source when you create the dashboard. In SourceEntity, you specify the type of object you're using as source. You can only create a dashboard from a template, so you use a SourceTemplate entity. If you need to create a dashboard from an analysis, first convert the analysis to a template by using the CreateTemplate API operation. For SourceTemplate, specify the Amazon Resource Name (ARN) of the source template. The SourceTemplateARN can contain any Amazon Web Services account and any Amazon QuickSight-supported Amazon Web Services Region.

Use the DataSetReferences entity within SourceTemplate to list the replacement datasets for the placeholders listed in the original. The schema in each dataset must match its placeholder.

Destructuring the Response

data CreateDashboardResponse

See: newCreateDashboardResponse smart constructor.

Constructors

CreateDashboardResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

arn :: Maybe Text

The ARN of the dashboard.

creationStatus :: Maybe ResourceStatus

The status of the dashboard creation request.

dashboardId :: Maybe Text

The ID for the dashboard.

versionArn :: Maybe Text

The ARN of the dashboard, including the version number of the first version that is created.

status :: Int

The HTTP status of the request.

newCreateDashboardResponse

Arguments

:: Int
$sel:status:CreateDashboardResponse'

-> CreateDashboardResponse

Create a value of CreateDashboardResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:CreateDashboardResponse', createDashboardResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:arn:CreateDashboardResponse', createDashboardResponse_arn - The ARN of the dashboard.

$sel:creationStatus:CreateDashboardResponse', createDashboardResponse_creationStatus - The status of the dashboard creation request.

$sel:dashboardId:CreateDashboard', createDashboardResponse_dashboardId - The ID for the dashboard.

$sel:versionArn:CreateDashboardResponse', createDashboardResponse_versionArn - The ARN of the dashboard, including the version number of the first version that is created.

$sel:status:CreateDashboardResponse', createDashboardResponse_status - The HTTP status of the request.

Response Lenses

createDashboardResponse_requestId :: Lens' CreateDashboardResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

createDashboardResponse_arn :: Lens' CreateDashboardResponse (Maybe Text)

The ARN of the dashboard.

createDashboardResponse_creationStatus :: Lens' CreateDashboardResponse (Maybe ResourceStatus)

The status of the dashboard creation request.

createDashboardResponse_dashboardId :: Lens' CreateDashboardResponse (Maybe Text)

The ID for the dashboard.

createDashboardResponse_versionArn :: Lens' CreateDashboardResponse (Maybe Text)

The ARN of the dashboard, including the version number of the first version that is created.

createDashboardResponse_status :: Lens' CreateDashboardResponse Int

The HTTP status of the request.