Network.AWS.QuickSight.CreateDataSource

Description

Creates a data source.

Creating a Request

data CreateDataSource

See: newCreateDataSource smart constructor.

Constructors

CreateDataSource'

Fields

dataSourceParameters :: Maybe DataSourceParameters

The parameters that Amazon QuickSight uses to connect to your underlying source.

sslProperties :: Maybe SslProperties

Secure Socket Layer (SSL) properties that apply when Amazon QuickSight connects to your underlying source.

credentials :: Maybe (Sensitive DataSourceCredentials)

The credentials Amazon QuickSight that uses to connect to your underlying source. Currently, only credentials based on user name and password are supported.

vpcConnectionProperties :: Maybe VpcConnectionProperties

Use this parameter only when you want Amazon QuickSight to use a VPC connection when connecting to your underlying source.

permissions :: Maybe (NonEmpty ResourcePermission)

A list of resource permissions on the data source.

tags :: Maybe (NonEmpty Tag)

Contains a map of the key-value pairs for the resource tag or tags assigned to the data source.

awsAccountId :: Text

The Amazon Web Services account ID.

dataSourceId :: Text

An ID for the data source. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

name :: Text

A display name for the data source.

type' :: DataSourceType

The type of the data source. To return a list of all data sources, use ListDataSources.

Use AMAZON_ELASTICSEARCH for Amazon OpenSearch Service.

newCreateDataSource

Arguments

:: Text
$sel:awsAccountId:CreateDataSource'

-> Text
$sel:dataSourceId:CreateDataSource'

-> Text
$sel:name:CreateDataSource'

-> DataSourceType
$sel:type':CreateDataSource'

-> CreateDataSource

Create a value of CreateDataSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataSourceParameters:CreateDataSource', createDataSource_dataSourceParameters - The parameters that Amazon QuickSight uses to connect to your underlying source.

$sel:sslProperties:CreateDataSource', createDataSource_sslProperties - Secure Socket Layer (SSL) properties that apply when Amazon QuickSight connects to your underlying source.

$sel:credentials:CreateDataSource', createDataSource_credentials - The credentials Amazon QuickSight that uses to connect to your underlying source. Currently, only credentials based on user name and password are supported.

$sel:vpcConnectionProperties:CreateDataSource', createDataSource_vpcConnectionProperties - Use this parameter only when you want Amazon QuickSight to use a VPC connection when connecting to your underlying source.

$sel:permissions:CreateDataSource', createDataSource_permissions - A list of resource permissions on the data source.

$sel:tags:CreateDataSource', createDataSource_tags - Contains a map of the key-value pairs for the resource tag or tags assigned to the data source.

$sel:awsAccountId:CreateDataSource', createDataSource_awsAccountId - The Amazon Web Services account ID.

$sel:dataSourceId:CreateDataSource', createDataSource_dataSourceId - An ID for the data source. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

$sel:name:CreateDataSource', createDataSource_name - A display name for the data source.

$sel:type':CreateDataSource', createDataSource_type - The type of the data source. To return a list of all data sources, use ListDataSources.

Use AMAZON_ELASTICSEARCH for Amazon OpenSearch Service.

Request Lenses

createDataSource_dataSourceParameters :: Lens' CreateDataSource (Maybe DataSourceParameters)

The parameters that Amazon QuickSight uses to connect to your underlying source.

createDataSource_sslProperties :: Lens' CreateDataSource (Maybe SslProperties)

Secure Socket Layer (SSL) properties that apply when Amazon QuickSight connects to your underlying source.

createDataSource_credentials :: Lens' CreateDataSource (Maybe DataSourceCredentials)

The credentials Amazon QuickSight that uses to connect to your underlying source. Currently, only credentials based on user name and password are supported.

createDataSource_vpcConnectionProperties :: Lens' CreateDataSource (Maybe VpcConnectionProperties)

Use this parameter only when you want Amazon QuickSight to use a VPC connection when connecting to your underlying source.

createDataSource_permissions :: Lens' CreateDataSource (Maybe (NonEmpty ResourcePermission))

A list of resource permissions on the data source.

createDataSource_tags :: Lens' CreateDataSource (Maybe (NonEmpty Tag))

Contains a map of the key-value pairs for the resource tag or tags assigned to the data source.

createDataSource_awsAccountId :: Lens' CreateDataSource Text

The Amazon Web Services account ID.

createDataSource_dataSourceId :: Lens' CreateDataSource Text

An ID for the data source. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

createDataSource_name :: Lens' CreateDataSource Text

A display name for the data source.

createDataSource_type :: Lens' CreateDataSource DataSourceType

The type of the data source. To return a list of all data sources, use ListDataSources.

Use AMAZON_ELASTICSEARCH for Amazon OpenSearch Service.

Destructuring the Response

data CreateDataSourceResponse

See: newCreateDataSourceResponse smart constructor.

Constructors

CreateDataSourceResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the data source.

creationStatus :: Maybe ResourceStatus

The status of creating the data source.

dataSourceId :: Maybe Text

The ID of the data source. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

status :: Int

The HTTP status of the request.

newCreateDataSourceResponse

Arguments

:: Int
$sel:status:CreateDataSourceResponse'

-> CreateDataSourceResponse

Create a value of CreateDataSourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:CreateDataSourceResponse', createDataSourceResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:arn:CreateDataSourceResponse', createDataSourceResponse_arn - The Amazon Resource Name (ARN) of the data source.

$sel:creationStatus:CreateDataSourceResponse', createDataSourceResponse_creationStatus - The status of creating the data source.

$sel:dataSourceId:CreateDataSource', createDataSourceResponse_dataSourceId - The ID of the data source. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

$sel:status:CreateDataSourceResponse', createDataSourceResponse_status - The HTTP status of the request.

Response Lenses

createDataSourceResponse_requestId :: Lens' CreateDataSourceResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

createDataSourceResponse_arn :: Lens' CreateDataSourceResponse (Maybe Text)

The Amazon Resource Name (ARN) of the data source.

createDataSourceResponse_creationStatus :: Lens' CreateDataSourceResponse (Maybe ResourceStatus)

The status of creating the data source.

createDataSourceResponse_dataSourceId :: Lens' CreateDataSourceResponse (Maybe Text)

The ID of the data source. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

createDataSourceResponse_status :: Lens' CreateDataSourceResponse Int

The HTTP status of the request.