Network.AWS.QuickSight.CreateGroup

Description

Creates an Amazon QuickSight group.

The permissions resource is arn:aws:quicksight:us-east-1:<relevant-aws-account-id>:group/default/<group-name> .

The response is a group object.

Creating a Request

data CreateGroup

The request object for this operation.

See: newCreateGroup smart constructor.

Constructors

CreateGroup'

Fields

description :: Maybe Text

A description for the group that you want to create.

groupName :: Text

A name for the group that you want to create.

awsAccountId :: Text

The ID for the Amazon Web Services account that the group is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

namespace :: Text

The namespace. Currently, you should set this to default.

newCreateGroup

Arguments

:: Text
$sel:groupName:CreateGroup'

-> Text
$sel:awsAccountId:CreateGroup'

-> Text
$sel:namespace:CreateGroup'

-> CreateGroup

Create a value of CreateGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:description:CreateGroup', createGroup_description - A description for the group that you want to create.

$sel:groupName:CreateGroup', createGroup_groupName - A name for the group that you want to create.

$sel:awsAccountId:CreateGroup', createGroup_awsAccountId - The ID for the Amazon Web Services account that the group is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

$sel:namespace:CreateGroup', createGroup_namespace - The namespace. Currently, you should set this to default.

Request Lenses

createGroup_description :: Lens' CreateGroup (Maybe Text)

A description for the group that you want to create.

createGroup_groupName :: Lens' CreateGroup Text

A name for the group that you want to create.

createGroup_awsAccountId :: Lens' CreateGroup Text

The ID for the Amazon Web Services account that the group is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

createGroup_namespace :: Lens' CreateGroup Text

The namespace. Currently, you should set this to default.

Destructuring the Response

data CreateGroupResponse

The response object for this operation.

See: newCreateGroupResponse smart constructor.

Constructors

CreateGroupResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

group' :: Maybe Group

The name of the group.

status :: Int

The HTTP status of the request.

newCreateGroupResponse

Arguments

:: Int
$sel:status:CreateGroupResponse'

-> CreateGroupResponse

Create a value of CreateGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:CreateGroupResponse', createGroupResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:group':CreateGroupResponse', createGroupResponse_group - The name of the group.

$sel:status:CreateGroupResponse', createGroupResponse_status - The HTTP status of the request.

Response Lenses

createGroupResponse_requestId :: Lens' CreateGroupResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

createGroupResponse_group :: Lens' CreateGroupResponse (Maybe Group)

The name of the group.

createGroupResponse_status :: Lens' CreateGroupResponse Int

The HTTP status of the request.