Network.AWS.QuickSight.CreateNamespace

Description

(Enterprise edition only) Creates a new namespace for you to use with Amazon QuickSight.

A namespace allows you to isolate the Amazon QuickSight users and groups that are registered for that namespace. Users that access the namespace can share assets only with other users or groups in the same namespace. They can't see users and groups in other namespaces. You can create a namespace after your Amazon Web Services account is subscribed to Amazon QuickSight. The namespace must be unique within the Amazon Web Services account. By default, there is a limit of 100 namespaces per Amazon Web Services account. To increase your limit, create a ticket with Amazon Web Services Support.

Creating a Request

data CreateNamespace

See: newCreateNamespace smart constructor.

Constructors

CreateNamespace'

Fields

tags :: Maybe (NonEmpty Tag)

The tags that you want to associate with the namespace that you're creating.

awsAccountId :: Text

The ID for the Amazon Web Services account that you want to create the Amazon QuickSight namespace in.

namespace :: Text

The name that you want to use to describe the new namespace.

identityStore :: IdentityStore

Specifies the type of your user identity directory. Currently, this supports users with an identity type of QUICKSIGHT.

newCreateNamespace

Arguments

:: Text
$sel:awsAccountId:CreateNamespace'

-> Text
$sel:namespace:CreateNamespace'

-> IdentityStore
$sel:identityStore:CreateNamespace'

-> CreateNamespace

Create a value of CreateNamespace with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateNamespace', createNamespace_tags - The tags that you want to associate with the namespace that you're creating.

$sel:awsAccountId:CreateNamespace', createNamespace_awsAccountId - The ID for the Amazon Web Services account that you want to create the Amazon QuickSight namespace in.

$sel:namespace:CreateNamespace', createNamespace_namespace - The name that you want to use to describe the new namespace.

$sel:identityStore:CreateNamespace', createNamespace_identityStore - Specifies the type of your user identity directory. Currently, this supports users with an identity type of QUICKSIGHT.

Request Lenses

createNamespace_tags :: Lens' CreateNamespace (Maybe (NonEmpty Tag))

The tags that you want to associate with the namespace that you're creating.

createNamespace_awsAccountId :: Lens' CreateNamespace Text

The ID for the Amazon Web Services account that you want to create the Amazon QuickSight namespace in.

createNamespace_namespace :: Lens' CreateNamespace Text

The name that you want to use to describe the new namespace.

createNamespace_identityStore :: Lens' CreateNamespace IdentityStore

Specifies the type of your user identity directory. Currently, this supports users with an identity type of QUICKSIGHT.

Destructuring the Response

data CreateNamespaceResponse

See: newCreateNamespaceResponse smart constructor.

Constructors

CreateNamespaceResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

capacityRegion :: Maybe Text

The Amazon Web Services Region; that you want to use for the free SPICE capacity for the new namespace. This is set to the region that you run CreateNamespace in.

arn :: Maybe Text

The ARN of the Amazon QuickSight namespace you created.

creationStatus :: Maybe NamespaceStatus

The status of the creation of the namespace. This is an asynchronous process. A status of CREATED means that your namespace is ready to use. If an error occurs, it indicates if the process is retryable or non-retryable. In the case of a non-retryable error, refer to the error message for follow-up tasks.

name :: Maybe Text

The name of the new namespace that you created.

identityStore :: Maybe IdentityStore

Specifies the type of your user identity directory. Currently, this supports users with an identity type of QUICKSIGHT.

status :: Int

The HTTP status of the request.

newCreateNamespaceResponse

Arguments

:: Int
$sel:status:CreateNamespaceResponse'

-> CreateNamespaceResponse

Create a value of CreateNamespaceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:CreateNamespaceResponse', createNamespaceResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:capacityRegion:CreateNamespaceResponse', createNamespaceResponse_capacityRegion - The Amazon Web Services Region; that you want to use for the free SPICE capacity for the new namespace. This is set to the region that you run CreateNamespace in.

$sel:arn:CreateNamespaceResponse', createNamespaceResponse_arn - The ARN of the Amazon QuickSight namespace you created.

$sel:creationStatus:CreateNamespaceResponse', createNamespaceResponse_creationStatus - The status of the creation of the namespace. This is an asynchronous process. A status of CREATED means that your namespace is ready to use. If an error occurs, it indicates if the process is retryable or non-retryable. In the case of a non-retryable error, refer to the error message for follow-up tasks.

$sel:name:CreateNamespaceResponse', createNamespaceResponse_name - The name of the new namespace that you created.

$sel:identityStore:CreateNamespace', createNamespaceResponse_identityStore - Specifies the type of your user identity directory. Currently, this supports users with an identity type of QUICKSIGHT.

$sel:status:CreateNamespaceResponse', createNamespaceResponse_status - The HTTP status of the request.

Response Lenses

createNamespaceResponse_requestId :: Lens' CreateNamespaceResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

createNamespaceResponse_capacityRegion :: Lens' CreateNamespaceResponse (Maybe Text)

The Amazon Web Services Region; that you want to use for the free SPICE capacity for the new namespace. This is set to the region that you run CreateNamespace in.

createNamespaceResponse_arn :: Lens' CreateNamespaceResponse (Maybe Text)

The ARN of the Amazon QuickSight namespace you created.

createNamespaceResponse_creationStatus :: Lens' CreateNamespaceResponse (Maybe NamespaceStatus)

The status of the creation of the namespace. This is an asynchronous process. A status of CREATED means that your namespace is ready to use. If an error occurs, it indicates if the process is retryable or non-retryable. In the case of a non-retryable error, refer to the error message for follow-up tasks.

createNamespaceResponse_name :: Lens' CreateNamespaceResponse (Maybe Text)

The name of the new namespace that you created.

createNamespaceResponse_identityStore :: Lens' CreateNamespaceResponse (Maybe IdentityStore)

Specifies the type of your user identity directory. Currently, this supports users with an identity type of QUICKSIGHT.

createNamespaceResponse_status :: Lens' CreateNamespaceResponse Int

The HTTP status of the request.