Network.AWS.QuickSight.CreateTheme

Description

Creates a theme.

A theme is set of configuration options for color and layout. Themes apply to analyses and dashboards. For more information, see Using Themes in Amazon QuickSight in the Amazon QuickSight User Guide.

Creating a Request

data CreateTheme

See: newCreateTheme smart constructor.

Constructors

CreateTheme'

Fields

versionDescription :: Maybe Text

A description of the first version of the theme that you're creating. Every time UpdateTheme is called, a new version is created. Each version of the theme has a description of the version in the VersionDescription field.

permissions :: Maybe (NonEmpty ResourcePermission)

A valid grouping of resource permissions to apply to the new theme.

tags :: Maybe (NonEmpty Tag)

A map of the key-value pairs for the resource tag or tags that you want to add to the resource.

awsAccountId :: Text

The ID of the Amazon Web Services account where you want to store the new theme.

themeId :: Text

An ID for the theme that you want to create. The theme ID is unique per Amazon Web Services Region in each Amazon Web Services account.

name :: Text

A display name for the theme.

baseThemeId :: Text

The ID of the theme that a custom theme will inherit from. All themes inherit from one of the starting themes defined by Amazon QuickSight. For a list of the starting themes, use ListThemes or choose Themes from within a Amazon QuickSight analysis.

configuration :: ThemeConfiguration

The theme configuration, which contains the theme display properties.

newCreateTheme

Arguments

:: Text
$sel:awsAccountId:CreateTheme'

-> Text
$sel:themeId:CreateTheme'

-> Text
$sel:name:CreateTheme'

-> Text
$sel:baseThemeId:CreateTheme'

-> ThemeConfiguration
$sel:configuration:CreateTheme'

-> CreateTheme

Create a value of CreateTheme with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:versionDescription:CreateTheme', createTheme_versionDescription - A description of the first version of the theme that you're creating. Every time UpdateTheme is called, a new version is created. Each version of the theme has a description of the version in the VersionDescription field.

$sel:permissions:CreateTheme', createTheme_permissions - A valid grouping of resource permissions to apply to the new theme.

$sel:tags:CreateTheme', createTheme_tags - A map of the key-value pairs for the resource tag or tags that you want to add to the resource.

$sel:awsAccountId:CreateTheme', createTheme_awsAccountId - The ID of the Amazon Web Services account where you want to store the new theme.

$sel:themeId:CreateTheme', createTheme_themeId - An ID for the theme that you want to create. The theme ID is unique per Amazon Web Services Region in each Amazon Web Services account.

$sel:name:CreateTheme', createTheme_name - A display name for the theme.

$sel:baseThemeId:CreateTheme', createTheme_baseThemeId - The ID of the theme that a custom theme will inherit from. All themes inherit from one of the starting themes defined by Amazon QuickSight. For a list of the starting themes, use ListThemes or choose Themes from within a Amazon QuickSight analysis.

$sel:configuration:CreateTheme', createTheme_configuration - The theme configuration, which contains the theme display properties.

Request Lenses

createTheme_versionDescription :: Lens' CreateTheme (Maybe Text)

A description of the first version of the theme that you're creating. Every time UpdateTheme is called, a new version is created. Each version of the theme has a description of the version in the VersionDescription field.

createTheme_permissions :: Lens' CreateTheme (Maybe (NonEmpty ResourcePermission))

A valid grouping of resource permissions to apply to the new theme.

createTheme_tags :: Lens' CreateTheme (Maybe (NonEmpty Tag))

A map of the key-value pairs for the resource tag or tags that you want to add to the resource.

createTheme_awsAccountId :: Lens' CreateTheme Text

The ID of the Amazon Web Services account where you want to store the new theme.

createTheme_themeId :: Lens' CreateTheme Text

An ID for the theme that you want to create. The theme ID is unique per Amazon Web Services Region in each Amazon Web Services account.

createTheme_name :: Lens' CreateTheme Text

A display name for the theme.

createTheme_baseThemeId :: Lens' CreateTheme Text

The ID of the theme that a custom theme will inherit from. All themes inherit from one of the starting themes defined by Amazon QuickSight. For a list of the starting themes, use ListThemes or choose Themes from within a Amazon QuickSight analysis.

createTheme_configuration :: Lens' CreateTheme ThemeConfiguration

The theme configuration, which contains the theme display properties.

Destructuring the Response

data CreateThemeResponse

See: newCreateThemeResponse smart constructor.

Constructors

CreateThemeResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

themeId :: Maybe Text

The ID of the theme.

arn :: Maybe Text

The Amazon Resource Name (ARN) for the theme.

creationStatus :: Maybe ResourceStatus

The theme creation status.

versionArn :: Maybe Text

The Amazon Resource Name (ARN) for the new theme.

status :: Int

The HTTP status of the request.

newCreateThemeResponse

Arguments

:: Int
$sel:status:CreateThemeResponse'

-> CreateThemeResponse

Create a value of CreateThemeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:CreateThemeResponse', createThemeResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:themeId:CreateTheme', createThemeResponse_themeId - The ID of the theme.

$sel:arn:CreateThemeResponse', createThemeResponse_arn - The Amazon Resource Name (ARN) for the theme.

$sel:creationStatus:CreateThemeResponse', createThemeResponse_creationStatus - The theme creation status.

$sel:versionArn:CreateThemeResponse', createThemeResponse_versionArn - The Amazon Resource Name (ARN) for the new theme.

$sel:status:CreateThemeResponse', createThemeResponse_status - The HTTP status of the request.

Response Lenses

createThemeResponse_requestId :: Lens' CreateThemeResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

createThemeResponse_themeId :: Lens' CreateThemeResponse (Maybe Text)

The ID of the theme.

createThemeResponse_arn :: Lens' CreateThemeResponse (Maybe Text)

The Amazon Resource Name (ARN) for the theme.

createThemeResponse_creationStatus :: Lens' CreateThemeResponse (Maybe ResourceStatus)

The theme creation status.

createThemeResponse_versionArn :: Lens' CreateThemeResponse (Maybe Text)

The Amazon Resource Name (ARN) for the new theme.

createThemeResponse_status :: Lens' CreateThemeResponse Int

The HTTP status of the request.