Network.AWS.QuickSight.DescribeAccountCustomization

Description

Describes the customizations associated with the provided Amazon Web Services account and Amazon Amazon QuickSight namespace in an Amazon Web Services Region. The Amazon QuickSight console evaluates which customizations to apply by running this API operation with the Resolved flag included.

To determine what customizations display when you run this command, it can help to visualize the relationship of the entities involved.

Amazon Web Services account - The Amazon Web Services account exists at the top of the hierarchy. It has the potential to use all of the Amazon Web Services Regions and AWS Services. When you subscribe to Amazon QuickSight, you choose one Amazon Web Services Region to use as your home Region. That's where your free SPICE capacity is located. You can use Amazon QuickSight in any supported Amazon Web Services Region.

Amazon Web Services Region - In each Amazon Web Services Region where you sign in to Amazon QuickSight at least once, Amazon QuickSight acts as a separate instance of the same service. If you have a user directory, it resides in us-east-1, which is the US East (N. Virginia). Generally speaking, these users have access to Amazon QuickSight in any Amazon Web Services Region, unless they are constrained to a namespace.

To run the command in a different Amazon Web Services Region, you change your Region settings. If you're using the AWS CLI, you can use one of the following options:

Use command line options.

Use named profiles.

Run aws configure to change your default Amazon Web Services Region. Use Enter to key the same settings for your keys. For more information, see Configuring the AWS CLI.

Namespace - A Amazon QuickSight namespace is a partition that contains users and assets (data sources, datasets, dashboards, and so on). To access assets that are in a specific namespace, users and groups must also be part of the same namespace. People who share a namespace are completely isolated from users and assets in other namespaces, even if they are in the same Amazon Web Services account and Amazon Web Services Region.

Applied customizations - Within an Amazon Web Services Region, a set of Amazon QuickSight customizations can apply to an Amazon Web Services account or to a namespace. Settings that you apply to a namespace override settings that you apply to an Amazon Web Services account. All settings are isolated to a single Amazon Web Services Region. To apply them in other Amazon Web Services Regions, run the CreateAccountCustomization command in each Amazon Web Services Region where you want to apply the same customizations.

Creating a Request

data DescribeAccountCustomization

See: newDescribeAccountCustomization smart constructor.

Constructors

DescribeAccountCustomization'

Fields

resolved :: Maybe Bool

The Resolved flag works with the other parameters to determine which view of Amazon QuickSight customizations is returned. You can add this flag to your command to use the same view that Amazon QuickSight uses to identify which customizations to apply to the console. Omit this flag, or set it to no-resolved, to reveal customizations that are configured at different levels.

namespace :: Maybe Text

The Amazon QuickSight namespace that you want to describe Amazon QuickSight customizations for.

awsAccountId :: Text

The ID for the Amazon Web Services account that you want to describe Amazon QuickSight customizations for.

newDescribeAccountCustomization

Arguments

:: Text
$sel:awsAccountId:DescribeAccountCustomization'

-> DescribeAccountCustomization

Create a value of DescribeAccountCustomization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resolved:DescribeAccountCustomization', describeAccountCustomization_resolved - The Resolved flag works with the other parameters to determine which view of Amazon QuickSight customizations is returned. You can add this flag to your command to use the same view that Amazon QuickSight uses to identify which customizations to apply to the console. Omit this flag, or set it to no-resolved, to reveal customizations that are configured at different levels.

$sel:namespace:DescribeAccountCustomization', describeAccountCustomization_namespace - The Amazon QuickSight namespace that you want to describe Amazon QuickSight customizations for.

$sel:awsAccountId:DescribeAccountCustomization', describeAccountCustomization_awsAccountId - The ID for the Amazon Web Services account that you want to describe Amazon QuickSight customizations for.

Request Lenses

describeAccountCustomization_resolved :: Lens' DescribeAccountCustomization (Maybe Bool)

The Resolved flag works with the other parameters to determine which view of Amazon QuickSight customizations is returned. You can add this flag to your command to use the same view that Amazon QuickSight uses to identify which customizations to apply to the console. Omit this flag, or set it to no-resolved, to reveal customizations that are configured at different levels.

describeAccountCustomization_namespace :: Lens' DescribeAccountCustomization (Maybe Text)

The Amazon QuickSight namespace that you want to describe Amazon QuickSight customizations for.

describeAccountCustomization_awsAccountId :: Lens' DescribeAccountCustomization Text

The ID for the Amazon Web Services account that you want to describe Amazon QuickSight customizations for.

Destructuring the Response

data DescribeAccountCustomizationResponse

See: newDescribeAccountCustomizationResponse smart constructor.

Constructors

DescribeAccountCustomizationResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

accountCustomization :: Maybe AccountCustomization

The Amazon QuickSight customizations that exist in the current Amazon Web Services Region.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the customization that's associated with this Amazon Web Services account.

namespace :: Maybe Text

The Amazon QuickSight namespace that you're describing.

awsAccountId :: Maybe Text

The ID for the Amazon Web Services account that you're describing.

status :: Int

The HTTP status of the request.

newDescribeAccountCustomizationResponse

Arguments

:: Int
$sel:status:DescribeAccountCustomizationResponse'

-> DescribeAccountCustomizationResponse

Create a value of DescribeAccountCustomizationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:DescribeAccountCustomizationResponse', describeAccountCustomizationResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:accountCustomization:DescribeAccountCustomizationResponse', describeAccountCustomizationResponse_accountCustomization - The Amazon QuickSight customizations that exist in the current Amazon Web Services Region.

$sel:arn:DescribeAccountCustomizationResponse', describeAccountCustomizationResponse_arn - The Amazon Resource Name (ARN) of the customization that's associated with this Amazon Web Services account.

$sel:namespace:DescribeAccountCustomization', describeAccountCustomizationResponse_namespace - The Amazon QuickSight namespace that you're describing.

$sel:awsAccountId:DescribeAccountCustomization', describeAccountCustomizationResponse_awsAccountId - The ID for the Amazon Web Services account that you're describing.

$sel:status:DescribeAccountCustomizationResponse', describeAccountCustomizationResponse_status - The HTTP status of the request.

Response Lenses

describeAccountCustomizationResponse_requestId :: Lens' DescribeAccountCustomizationResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

describeAccountCustomizationResponse_accountCustomization :: Lens' DescribeAccountCustomizationResponse (Maybe AccountCustomization)

The Amazon QuickSight customizations that exist in the current Amazon Web Services Region.

describeAccountCustomizationResponse_arn :: Lens' DescribeAccountCustomizationResponse (Maybe Text)

The Amazon Resource Name (ARN) of the customization that's associated with this Amazon Web Services account.

describeAccountCustomizationResponse_namespace :: Lens' DescribeAccountCustomizationResponse (Maybe Text)

The Amazon QuickSight namespace that you're describing.

describeAccountCustomizationResponse_awsAccountId :: Lens' DescribeAccountCustomizationResponse (Maybe Text)

The ID for the Amazon Web Services account that you're describing.

describeAccountCustomizationResponse_status :: Lens' DescribeAccountCustomizationResponse Int

The HTTP status of the request.