Network.AWS.QuickSight.DescribeAnalysisPermissions

Description

Provides the read and write permissions for an analysis.

Creating a Request

data DescribeAnalysisPermissions

See: newDescribeAnalysisPermissions smart constructor.

Constructors

DescribeAnalysisPermissions'

Fields

awsAccountId :: Text

The ID of the Amazon Web Services account that contains the analysis whose permissions you're describing. You must be using the Amazon Web Services account that the analysis is in.

analysisId :: Text

The ID of the analysis whose permissions you're describing. The ID is part of the analysis URL.

newDescribeAnalysisPermissions

Arguments

:: Text
$sel:awsAccountId:DescribeAnalysisPermissions'

-> Text
$sel:analysisId:DescribeAnalysisPermissions'

-> DescribeAnalysisPermissions

Create a value of DescribeAnalysisPermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsAccountId:DescribeAnalysisPermissions', describeAnalysisPermissions_awsAccountId - The ID of the Amazon Web Services account that contains the analysis whose permissions you're describing. You must be using the Amazon Web Services account that the analysis is in.

$sel:analysisId:DescribeAnalysisPermissions', describeAnalysisPermissions_analysisId - The ID of the analysis whose permissions you're describing. The ID is part of the analysis URL.

Request Lenses

describeAnalysisPermissions_awsAccountId :: Lens' DescribeAnalysisPermissions Text

The ID of the Amazon Web Services account that contains the analysis whose permissions you're describing. You must be using the Amazon Web Services account that the analysis is in.

describeAnalysisPermissions_analysisId :: Lens' DescribeAnalysisPermissions Text

The ID of the analysis whose permissions you're describing. The ID is part of the analysis URL.

Destructuring the Response

data DescribeAnalysisPermissionsResponse

See: newDescribeAnalysisPermissionsResponse smart constructor.

Constructors

DescribeAnalysisPermissionsResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

analysisId :: Maybe Text

The ID of the analysis whose permissions you're describing.

analysisArn :: Maybe Text

The Amazon Resource Name (ARN) of the analysis whose permissions you're describing.

permissions :: Maybe (NonEmpty ResourcePermission)

A structure that describes the principals and the resource-level permissions on an analysis.

status :: Int

The HTTP status of the request.

newDescribeAnalysisPermissionsResponse

Arguments

:: Int
$sel:status:DescribeAnalysisPermissionsResponse'

-> DescribeAnalysisPermissionsResponse

Create a value of DescribeAnalysisPermissionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:DescribeAnalysisPermissionsResponse', describeAnalysisPermissionsResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:analysisId:DescribeAnalysisPermissions', describeAnalysisPermissionsResponse_analysisId - The ID of the analysis whose permissions you're describing.

$sel:analysisArn:DescribeAnalysisPermissionsResponse', describeAnalysisPermissionsResponse_analysisArn - The Amazon Resource Name (ARN) of the analysis whose permissions you're describing.

$sel:permissions:DescribeAnalysisPermissionsResponse', describeAnalysisPermissionsResponse_permissions - A structure that describes the principals and the resource-level permissions on an analysis.

$sel:status:DescribeAnalysisPermissionsResponse', describeAnalysisPermissionsResponse_status - The HTTP status of the request.

Response Lenses

describeAnalysisPermissionsResponse_requestId :: Lens' DescribeAnalysisPermissionsResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

describeAnalysisPermissionsResponse_analysisId :: Lens' DescribeAnalysisPermissionsResponse (Maybe Text)

The ID of the analysis whose permissions you're describing.

describeAnalysisPermissionsResponse_analysisArn :: Lens' DescribeAnalysisPermissionsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the analysis whose permissions you're describing.

describeAnalysisPermissionsResponse_permissions :: Lens' DescribeAnalysisPermissionsResponse (Maybe (NonEmpty ResourcePermission))

A structure that describes the principals and the resource-level permissions on an analysis.

describeAnalysisPermissionsResponse_status :: Lens' DescribeAnalysisPermissionsResponse Int

The HTTP status of the request.