Network.AWS.QuickSight.DescribeDataSource

Description

Describes a data source.

Creating a Request

data DescribeDataSource

See: newDescribeDataSource smart constructor.

Constructors

DescribeDataSource'

Fields

awsAccountId :: Text

The Amazon Web Services account ID.

dataSourceId :: Text

The ID of the data source. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

newDescribeDataSource

Arguments

:: Text
$sel:awsAccountId:DescribeDataSource'

-> Text
$sel:dataSourceId:DescribeDataSource'

-> DescribeDataSource

Create a value of DescribeDataSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsAccountId:DescribeDataSource', describeDataSource_awsAccountId - The Amazon Web Services account ID.

$sel:dataSourceId:DescribeDataSource', describeDataSource_dataSourceId - The ID of the data source. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

Request Lenses

describeDataSource_awsAccountId :: Lens' DescribeDataSource Text

The Amazon Web Services account ID.

describeDataSource_dataSourceId :: Lens' DescribeDataSource Text

The ID of the data source. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

Destructuring the Response

data DescribeDataSourceResponse

See: newDescribeDataSourceResponse smart constructor.

Constructors

DescribeDataSourceResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

dataSource :: Maybe DataSource

The information on the data source.

status :: Int

The HTTP status of the request.

newDescribeDataSourceResponse

Arguments

:: Int
$sel:status:DescribeDataSourceResponse'

-> DescribeDataSourceResponse

Create a value of DescribeDataSourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:DescribeDataSourceResponse', describeDataSourceResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:dataSource:DescribeDataSourceResponse', describeDataSourceResponse_dataSource - The information on the data source.

$sel:status:DescribeDataSourceResponse', describeDataSourceResponse_status - The HTTP status of the request.

Response Lenses

describeDataSourceResponse_requestId :: Lens' DescribeDataSourceResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

describeDataSourceResponse_dataSource :: Lens' DescribeDataSourceResponse (Maybe DataSource)

The information on the data source.

describeDataSourceResponse_status :: Lens' DescribeDataSourceResponse Int

The HTTP status of the request.