Network.AWS.QuickSight.DescribeTemplate

Description

Describes a template's metadata.

Creating a Request

data DescribeTemplate

See: newDescribeTemplate smart constructor.

Constructors

DescribeTemplate'

Fields

aliasName :: Maybe Text

The alias of the template that you want to describe. If you name a specific alias, you describe the version that the alias points to. You can specify the latest version of the template by providing the keyword $LATEST in the AliasName parameter. The keyword $PUBLISHED doesn't apply to templates.

versionNumber :: Maybe Natural

(Optional) The number for the version to describe. If a VersionNumber parameter value isn't provided, the latest version of the template is described.

awsAccountId :: Text

The ID of the Amazon Web Services account that contains the template that you're describing.

templateId :: Text

The ID for the template.

newDescribeTemplate

Arguments

:: Text
$sel:awsAccountId:DescribeTemplate'

-> Text
$sel:templateId:DescribeTemplate'

-> DescribeTemplate

Create a value of DescribeTemplate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:aliasName:DescribeTemplate', describeTemplate_aliasName - The alias of the template that you want to describe. If you name a specific alias, you describe the version that the alias points to. You can specify the latest version of the template by providing the keyword $LATEST in the AliasName parameter. The keyword $PUBLISHED doesn't apply to templates.

$sel:versionNumber:DescribeTemplate', describeTemplate_versionNumber - (Optional) The number for the version to describe. If a VersionNumber parameter value isn't provided, the latest version of the template is described.

$sel:awsAccountId:DescribeTemplate', describeTemplate_awsAccountId - The ID of the Amazon Web Services account that contains the template that you're describing.

$sel:templateId:DescribeTemplate', describeTemplate_templateId - The ID for the template.

Request Lenses

describeTemplate_aliasName :: Lens' DescribeTemplate (Maybe Text)

The alias of the template that you want to describe. If you name a specific alias, you describe the version that the alias points to. You can specify the latest version of the template by providing the keyword $LATEST in the AliasName parameter. The keyword $PUBLISHED doesn't apply to templates.

describeTemplate_versionNumber :: Lens' DescribeTemplate (Maybe Natural)

(Optional) The number for the version to describe. If a VersionNumber parameter value isn't provided, the latest version of the template is described.

describeTemplate_awsAccountId :: Lens' DescribeTemplate Text

The ID of the Amazon Web Services account that contains the template that you're describing.

describeTemplate_templateId :: Lens' DescribeTemplate Text

The ID for the template.

Destructuring the Response

data DescribeTemplateResponse

See: newDescribeTemplateResponse smart constructor.

Constructors

DescribeTemplateResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

template :: Maybe Template

The template structure for the object you want to describe.

status :: Int

The HTTP status of the request.

newDescribeTemplateResponse

Arguments

:: Int
$sel:status:DescribeTemplateResponse'

-> DescribeTemplateResponse

Create a value of DescribeTemplateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:DescribeTemplateResponse', describeTemplateResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:template:DescribeTemplateResponse', describeTemplateResponse_template - The template structure for the object you want to describe.

$sel:status:DescribeTemplateResponse', describeTemplateResponse_status - The HTTP status of the request.

Response Lenses

describeTemplateResponse_requestId :: Lens' DescribeTemplateResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

describeTemplateResponse_template :: Lens' DescribeTemplateResponse (Maybe Template)

The template structure for the object you want to describe.

describeTemplateResponse_status :: Lens' DescribeTemplateResponse Int

The HTTP status of the request.