Network.AWS.QuickSight.DescribeThemePermissions

Description

Describes the read and write permissions for a theme.

Creating a Request

data DescribeThemePermissions

See: newDescribeThemePermissions smart constructor.

Constructors

DescribeThemePermissions'

Fields

awsAccountId :: Text

The ID of the Amazon Web Services account that contains the theme that you're describing.

themeId :: Text

The ID for the theme that you want to describe permissions for.

newDescribeThemePermissions

Arguments

:: Text
$sel:awsAccountId:DescribeThemePermissions'

-> Text
$sel:themeId:DescribeThemePermissions'

-> DescribeThemePermissions

Create a value of DescribeThemePermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsAccountId:DescribeThemePermissions', describeThemePermissions_awsAccountId - The ID of the Amazon Web Services account that contains the theme that you're describing.

$sel:themeId:DescribeThemePermissions', describeThemePermissions_themeId - The ID for the theme that you want to describe permissions for.

Request Lenses

describeThemePermissions_awsAccountId :: Lens' DescribeThemePermissions Text

The ID of the Amazon Web Services account that contains the theme that you're describing.

describeThemePermissions_themeId :: Lens' DescribeThemePermissions Text

The ID for the theme that you want to describe permissions for.

Destructuring the Response

data DescribeThemePermissionsResponse

See: newDescribeThemePermissionsResponse smart constructor.

Constructors

DescribeThemePermissionsResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

themeArn :: Maybe Text

The Amazon Resource Name (ARN) of the theme.

themeId :: Maybe Text

The ID for the theme.

permissions :: Maybe (NonEmpty ResourcePermission)

A list of resource permissions set on the theme.

status :: Int

The HTTP status of the request.

newDescribeThemePermissionsResponse

Arguments

:: Int
$sel:status:DescribeThemePermissionsResponse'

-> DescribeThemePermissionsResponse

Create a value of DescribeThemePermissionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:DescribeThemePermissionsResponse', describeThemePermissionsResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:themeArn:DescribeThemePermissionsResponse', describeThemePermissionsResponse_themeArn - The Amazon Resource Name (ARN) of the theme.

$sel:themeId:DescribeThemePermissions', describeThemePermissionsResponse_themeId - The ID for the theme.

$sel:permissions:DescribeThemePermissionsResponse', describeThemePermissionsResponse_permissions - A list of resource permissions set on the theme.

$sel:status:DescribeThemePermissionsResponse', describeThemePermissionsResponse_status - The HTTP status of the request.

Response Lenses

describeThemePermissionsResponse_requestId :: Lens' DescribeThemePermissionsResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

describeThemePermissionsResponse_themeArn :: Lens' DescribeThemePermissionsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the theme.

describeThemePermissionsResponse_themeId :: Lens' DescribeThemePermissionsResponse (Maybe Text)

The ID for the theme.

describeThemePermissionsResponse_permissions :: Lens' DescribeThemePermissionsResponse (Maybe (NonEmpty ResourcePermission))

A list of resource permissions set on the theme.

describeThemePermissionsResponse_status :: Lens' DescribeThemePermissionsResponse Int

The HTTP status of the request.