Network.AWS.QuickSight.DescribeUser

Description

Returns information about a user, given the user name.

Creating a Request

data DescribeUser

See: newDescribeUser smart constructor.

Constructors

DescribeUser'

Fields

userName :: Text

The name of the user that you want to describe.

awsAccountId :: Text

The ID for the Amazon Web Services account that the user is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

namespace :: Text

The namespace. Currently, you should set this to default.

newDescribeUser

Arguments

:: Text
$sel:userName:DescribeUser'

-> Text
$sel:awsAccountId:DescribeUser'

-> Text
$sel:namespace:DescribeUser'

-> DescribeUser

Create a value of DescribeUser with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:userName:DescribeUser', describeUser_userName - The name of the user that you want to describe.

$sel:awsAccountId:DescribeUser', describeUser_awsAccountId - The ID for the Amazon Web Services account that the user is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

$sel:namespace:DescribeUser', describeUser_namespace - The namespace. Currently, you should set this to default.

Request Lenses

describeUser_userName :: Lens' DescribeUser Text

The name of the user that you want to describe.

describeUser_awsAccountId :: Lens' DescribeUser Text

The ID for the Amazon Web Services account that the user is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

describeUser_namespace :: Lens' DescribeUser Text

The namespace. Currently, you should set this to default.

Destructuring the Response

data DescribeUserResponse

See: newDescribeUserResponse smart constructor.

Constructors

DescribeUserResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

user :: Maybe User

The user name.

status :: Int

The HTTP status of the request.

newDescribeUserResponse

Arguments

:: Int
$sel:status:DescribeUserResponse'

-> DescribeUserResponse

Create a value of DescribeUserResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:DescribeUserResponse', describeUserResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:user:DescribeUserResponse', describeUserResponse_user - The user name.

$sel:status:DescribeUserResponse', describeUserResponse_status - The HTTP status of the request.

Response Lenses

describeUserResponse_requestId :: Lens' DescribeUserResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

describeUserResponse_user :: Lens' DescribeUserResponse (Maybe User)

The user name.

describeUserResponse_status :: Lens' DescribeUserResponse Int

The HTTP status of the request.