Network.AWS.QuickSight.GetSessionEmbedUrl

Description

Generates a session URL and authorization code that you can use to embed the Amazon Amazon QuickSight console in your web server code. Use GetSessionEmbedUrl where you want to provide an authoring portal that allows users to create data sources, datasets, analyses, and dashboards. The users who access an embedded Amazon QuickSight console need belong to the author or admin security cohort. If you want to restrict permissions to some of these features, add a custom permissions profile to the user with the UpdateUser API operation. Use RegisterUser API operation to add a new user with a custom permission profile attached. For more information, see the following sections in the Amazon QuickSight User Guide:

Embedding Analytics

Customizing Access to the Amazon QuickSight Console

Creating a Request

data GetSessionEmbedUrl

See: newGetSessionEmbedUrl smart constructor.

Constructors

GetSessionEmbedUrl'

Fields

sessionLifetimeInMinutes :: Maybe Natural

How many minutes the session is valid. The session lifetime must be 15-600 minutes.

entryPoint :: Maybe Text

The URL you use to access the embedded session. The entry point URL is constrained to the following paths:

/start

/start/analyses

/start/dashboards

/start/favorites

/dashboards/DashboardId - where DashboardId is the actual ID key from the Amazon QuickSight console URL of the dashboard

/analyses/AnalysisId - where AnalysisId is the actual ID key from the Amazon QuickSight console URL of the analysis

userArn :: Maybe Text

The Amazon QuickSight user's Amazon Resource Name (ARN), for use with QUICKSIGHT identity type. You can use this for any type of Amazon QuickSight users in your account (readers, authors, or admins). They need to be authenticated as one of the following:

Active Directory (AD) users or group members

Invited nonfederated users

Identity and Access Management (IAM) users and IAM role-based sessions authenticated through Federated Single Sign-On using SAML, OpenID Connect, or IAM federation

Omit this parameter for users in the third group, IAM users and IAM role-based sessions.

awsAccountId :: Text

The ID for the Amazon Web Services account associated with your Amazon QuickSight subscription.

newGetSessionEmbedUrl

Arguments

:: Text
$sel:awsAccountId:GetSessionEmbedUrl'

-> GetSessionEmbedUrl

Create a value of GetSessionEmbedUrl with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sessionLifetimeInMinutes:GetSessionEmbedUrl', getSessionEmbedUrl_sessionLifetimeInMinutes - How many minutes the session is valid. The session lifetime must be 15-600 minutes.

$sel:entryPoint:GetSessionEmbedUrl', getSessionEmbedUrl_entryPoint - The URL you use to access the embedded session. The entry point URL is constrained to the following paths:

/start

/start/analyses

/start/dashboards

/start/favorites

/dashboards/DashboardId - where DashboardId is the actual ID key from the Amazon QuickSight console URL of the dashboard

/analyses/AnalysisId - where AnalysisId is the actual ID key from the Amazon QuickSight console URL of the analysis

$sel:userArn:GetSessionEmbedUrl', getSessionEmbedUrl_userArn - The Amazon QuickSight user's Amazon Resource Name (ARN), for use with QUICKSIGHT identity type. You can use this for any type of Amazon QuickSight users in your account (readers, authors, or admins). They need to be authenticated as one of the following:

Active Directory (AD) users or group members

Invited nonfederated users

Identity and Access Management (IAM) users and IAM role-based sessions authenticated through Federated Single Sign-On using SAML, OpenID Connect, or IAM federation

Omit this parameter for users in the third group, IAM users and IAM role-based sessions.

$sel:awsAccountId:GetSessionEmbedUrl', getSessionEmbedUrl_awsAccountId - The ID for the Amazon Web Services account associated with your Amazon QuickSight subscription.

Request Lenses

getSessionEmbedUrl_sessionLifetimeInMinutes :: Lens' GetSessionEmbedUrl (Maybe Natural)

How many minutes the session is valid. The session lifetime must be 15-600 minutes.

getSessionEmbedUrl_entryPoint :: Lens' GetSessionEmbedUrl (Maybe Text)

The URL you use to access the embedded session. The entry point URL is constrained to the following paths:

/start

/start/analyses

/start/dashboards

/start/favorites

/dashboards/DashboardId - where DashboardId is the actual ID key from the Amazon QuickSight console URL of the dashboard

/analyses/AnalysisId - where AnalysisId is the actual ID key from the Amazon QuickSight console URL of the analysis

getSessionEmbedUrl_userArn :: Lens' GetSessionEmbedUrl (Maybe Text)

The Amazon QuickSight user's Amazon Resource Name (ARN), for use with QUICKSIGHT identity type. You can use this for any type of Amazon QuickSight users in your account (readers, authors, or admins). They need to be authenticated as one of the following:

Active Directory (AD) users or group members

Invited nonfederated users

Identity and Access Management (IAM) users and IAM role-based sessions authenticated through Federated Single Sign-On using SAML, OpenID Connect, or IAM federation

Omit this parameter for users in the third group, IAM users and IAM role-based sessions.

getSessionEmbedUrl_awsAccountId :: Lens' GetSessionEmbedUrl Text

The ID for the Amazon Web Services account associated with your Amazon QuickSight subscription.

Destructuring the Response

data GetSessionEmbedUrlResponse

See: newGetSessionEmbedUrlResponse smart constructor.

Constructors

GetSessionEmbedUrlResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

embedUrl :: Maybe (Sensitive Text)

A single-use URL that you can put into your server-side web page to embed your Amazon QuickSight session. This URL is valid for 5 minutes. The API operation provides the URL with an auth_code value that enables one (and only one) sign-on to a user session that is valid for 10 hours.

status :: Int

The HTTP status of the request.

newGetSessionEmbedUrlResponse

Arguments

:: Int
$sel:status:GetSessionEmbedUrlResponse'

-> GetSessionEmbedUrlResponse

Create a value of GetSessionEmbedUrlResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:GetSessionEmbedUrlResponse', getSessionEmbedUrlResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:embedUrl:GetSessionEmbedUrlResponse', getSessionEmbedUrlResponse_embedUrl - A single-use URL that you can put into your server-side web page to embed your Amazon QuickSight session. This URL is valid for 5 minutes. The API operation provides the URL with an auth_code value that enables one (and only one) sign-on to a user session that is valid for 10 hours.

$sel:status:GetSessionEmbedUrlResponse', getSessionEmbedUrlResponse_status - The HTTP status of the request.

Response Lenses

getSessionEmbedUrlResponse_requestId :: Lens' GetSessionEmbedUrlResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

getSessionEmbedUrlResponse_embedUrl :: Lens' GetSessionEmbedUrlResponse (Maybe Text)

A single-use URL that you can put into your server-side web page to embed your Amazon QuickSight session. This URL is valid for 5 minutes. The API operation provides the URL with an auth_code value that enables one (and only one) sign-on to a user session that is valid for 10 hours.

getSessionEmbedUrlResponse_status :: Lens' GetSessionEmbedUrlResponse Int

The HTTP status of the request.