Network.AWS.QuickSight.ListGroups

Description

Lists all user groups in Amazon QuickSight.

Creating a Request

data ListGroups

See: newListGroups smart constructor.

Constructors

ListGroups'

Fields

nextToken :: Maybe Text

A pagination token that can be used in a subsequent request.

maxResults :: Maybe Natural

The maximum number of results to return.

awsAccountId :: Text

The ID for the Amazon Web Services account that the group is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

namespace :: Text

The namespace. Currently, you should set this to default.

newListGroups

Arguments

:: Text
$sel:awsAccountId:ListGroups'

-> Text
$sel:namespace:ListGroups'

-> ListGroups

Create a value of ListGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListGroups', listGroups_nextToken - A pagination token that can be used in a subsequent request.

$sel:maxResults:ListGroups', listGroups_maxResults - The maximum number of results to return.

$sel:awsAccountId:ListGroups', listGroups_awsAccountId - The ID for the Amazon Web Services account that the group is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

$sel:namespace:ListGroups', listGroups_namespace - The namespace. Currently, you should set this to default.

Request Lenses

listGroups_nextToken :: Lens' ListGroups (Maybe Text)

A pagination token that can be used in a subsequent request.

listGroups_maxResults :: Lens' ListGroups (Maybe Natural)

The maximum number of results to return.

listGroups_awsAccountId :: Lens' ListGroups Text

The ID for the Amazon Web Services account that the group is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

listGroups_namespace :: Lens' ListGroups Text

The namespace. Currently, you should set this to default.

Destructuring the Response

data ListGroupsResponse

See: newListGroupsResponse smart constructor.

Constructors

ListGroupsResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

groupList :: Maybe [Group]

The list of the groups.

nextToken :: Maybe Text

A pagination token that can be used in a subsequent request.

status :: Int

The HTTP status of the request.

newListGroupsResponse

Arguments

:: Int
$sel:status:ListGroupsResponse'

-> ListGroupsResponse

Create a value of ListGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:ListGroupsResponse', listGroupsResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:groupList:ListGroupsResponse', listGroupsResponse_groupList - The list of the groups.

$sel:nextToken:ListGroups', listGroupsResponse_nextToken - A pagination token that can be used in a subsequent request.

$sel:status:ListGroupsResponse', listGroupsResponse_status - The HTTP status of the request.

Response Lenses

listGroupsResponse_requestId :: Lens' ListGroupsResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

listGroupsResponse_groupList :: Lens' ListGroupsResponse (Maybe [Group])

The list of the groups.

listGroupsResponse_nextToken :: Lens' ListGroupsResponse (Maybe Text)

A pagination token that can be used in a subsequent request.

listGroupsResponse_status :: Lens' ListGroupsResponse Int

The HTTP status of the request.