Network.AWS.QuickSight.ListTemplateVersions

Description

Lists all the versions of the templates in the current Amazon QuickSight account.

This operation returns paginated results.

Creating a Request

data ListTemplateVersions

See: newListTemplateVersions smart constructor.

Constructors

ListTemplateVersions'

Fields

nextToken :: Maybe Text

The token for the next set of results, or null if there are no more results.

maxResults :: Maybe Natural

The maximum number of results to be returned per request.

awsAccountId :: Text

The ID of the Amazon Web Services account that contains the templates that you're listing.

templateId :: Text

The ID for the template.

newListTemplateVersions

Arguments

:: Text
$sel:awsAccountId:ListTemplateVersions'

-> Text
$sel:templateId:ListTemplateVersions'

-> ListTemplateVersions

Create a value of ListTemplateVersions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListTemplateVersions', listTemplateVersions_nextToken - The token for the next set of results, or null if there are no more results.

$sel:maxResults:ListTemplateVersions', listTemplateVersions_maxResults - The maximum number of results to be returned per request.

$sel:awsAccountId:ListTemplateVersions', listTemplateVersions_awsAccountId - The ID of the Amazon Web Services account that contains the templates that you're listing.

$sel:templateId:ListTemplateVersions', listTemplateVersions_templateId - The ID for the template.

Request Lenses

listTemplateVersions_nextToken :: Lens' ListTemplateVersions (Maybe Text)

The token for the next set of results, or null if there are no more results.

listTemplateVersions_maxResults :: Lens' ListTemplateVersions (Maybe Natural)

The maximum number of results to be returned per request.

listTemplateVersions_awsAccountId :: Lens' ListTemplateVersions Text

The ID of the Amazon Web Services account that contains the templates that you're listing.

listTemplateVersions_templateId :: Lens' ListTemplateVersions Text

The ID for the template.

Destructuring the Response

data ListTemplateVersionsResponse

See: newListTemplateVersionsResponse smart constructor.

Constructors

ListTemplateVersionsResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

templateVersionSummaryList :: Maybe [TemplateVersionSummary]

A structure containing a list of all the versions of the specified template.

nextToken :: Maybe Text

The token for the next set of results, or null if there are no more results.

status :: Int

The HTTP status of the request.

newListTemplateVersionsResponse

Arguments

:: Int
$sel:status:ListTemplateVersionsResponse'

-> ListTemplateVersionsResponse

Create a value of ListTemplateVersionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:ListTemplateVersionsResponse', listTemplateVersionsResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:templateVersionSummaryList:ListTemplateVersionsResponse', listTemplateVersionsResponse_templateVersionSummaryList - A structure containing a list of all the versions of the specified template.

$sel:nextToken:ListTemplateVersions', listTemplateVersionsResponse_nextToken - The token for the next set of results, or null if there are no more results.

$sel:status:ListTemplateVersionsResponse', listTemplateVersionsResponse_status - The HTTP status of the request.

Response Lenses

listTemplateVersionsResponse_requestId :: Lens' ListTemplateVersionsResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

listTemplateVersionsResponse_templateVersionSummaryList :: Lens' ListTemplateVersionsResponse (Maybe [TemplateVersionSummary])

A structure containing a list of all the versions of the specified template.

listTemplateVersionsResponse_nextToken :: Lens' ListTemplateVersionsResponse (Maybe Text)

The token for the next set of results, or null if there are no more results.

listTemplateVersionsResponse_status :: Lens' ListTemplateVersionsResponse Int

The HTTP status of the request.