Network.AWS.QuickSight.ListUserGroups

Description

Lists the Amazon QuickSight groups that an Amazon QuickSight user is a member of.

Creating a Request

data ListUserGroups

See: newListUserGroups smart constructor.

Constructors

ListUserGroups'

Fields

nextToken :: Maybe Text

A pagination token that can be used in a subsequent request.

maxResults :: Maybe Natural

The maximum number of results to return from this request.

userName :: Text

The Amazon QuickSight user name that you want to list group memberships for.

awsAccountId :: Text

The Amazon Web Services account ID that the user is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

namespace :: Text

The namespace. Currently, you should set this to default.

newListUserGroups

Arguments

:: Text
$sel:userName:ListUserGroups'

-> Text
$sel:awsAccountId:ListUserGroups'

-> Text
$sel:namespace:ListUserGroups'

-> ListUserGroups

Create a value of ListUserGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:ListUserGroups', listUserGroups_nextToken - A pagination token that can be used in a subsequent request.

$sel:maxResults:ListUserGroups', listUserGroups_maxResults - The maximum number of results to return from this request.

$sel:userName:ListUserGroups', listUserGroups_userName - The Amazon QuickSight user name that you want to list group memberships for.

$sel:awsAccountId:ListUserGroups', listUserGroups_awsAccountId - The Amazon Web Services account ID that the user is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

$sel:namespace:ListUserGroups', listUserGroups_namespace - The namespace. Currently, you should set this to default.

Request Lenses

listUserGroups_nextToken :: Lens' ListUserGroups (Maybe Text)

A pagination token that can be used in a subsequent request.

listUserGroups_maxResults :: Lens' ListUserGroups (Maybe Natural)

The maximum number of results to return from this request.

listUserGroups_userName :: Lens' ListUserGroups Text

The Amazon QuickSight user name that you want to list group memberships for.

listUserGroups_awsAccountId :: Lens' ListUserGroups Text

The Amazon Web Services account ID that the user is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

listUserGroups_namespace :: Lens' ListUserGroups Text

The namespace. Currently, you should set this to default.

Destructuring the Response

data ListUserGroupsResponse

See: newListUserGroupsResponse smart constructor.

Constructors

ListUserGroupsResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

groupList :: Maybe [Group]

The list of groups the user is a member of.

nextToken :: Maybe Text

A pagination token that can be used in a subsequent request.

status :: Int

The HTTP status of the request.

newListUserGroupsResponse

Arguments

:: Int
$sel:status:ListUserGroupsResponse'

-> ListUserGroupsResponse

Create a value of ListUserGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:ListUserGroupsResponse', listUserGroupsResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:groupList:ListUserGroupsResponse', listUserGroupsResponse_groupList - The list of groups the user is a member of.

$sel:nextToken:ListUserGroups', listUserGroupsResponse_nextToken - A pagination token that can be used in a subsequent request.

$sel:status:ListUserGroupsResponse', listUserGroupsResponse_status - The HTTP status of the request.

Response Lenses

listUserGroupsResponse_requestId :: Lens' ListUserGroupsResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

listUserGroupsResponse_groupList :: Lens' ListUserGroupsResponse (Maybe [Group])

The list of groups the user is a member of.

listUserGroupsResponse_nextToken :: Lens' ListUserGroupsResponse (Maybe Text)

A pagination token that can be used in a subsequent request.

listUserGroupsResponse_status :: Lens' ListUserGroupsResponse Int

The HTTP status of the request.