Network.AWS.QuickSight.RestoreAnalysis

Description

Restores an analysis.

Creating a Request

data RestoreAnalysis

See: newRestoreAnalysis smart constructor.

Constructors

RestoreAnalysis'

Fields

awsAccountId :: Text

The ID of the Amazon Web Services account that contains the analysis.

analysisId :: Text

The ID of the analysis that you're restoring.

newRestoreAnalysis

Arguments

:: Text
$sel:awsAccountId:RestoreAnalysis'

-> Text
$sel:analysisId:RestoreAnalysis'

-> RestoreAnalysis

Create a value of RestoreAnalysis with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:awsAccountId:RestoreAnalysis', restoreAnalysis_awsAccountId - The ID of the Amazon Web Services account that contains the analysis.

$sel:analysisId:RestoreAnalysis', restoreAnalysis_analysisId - The ID of the analysis that you're restoring.

Request Lenses

restoreAnalysis_awsAccountId :: Lens' RestoreAnalysis Text

The ID of the Amazon Web Services account that contains the analysis.

restoreAnalysis_analysisId :: Lens' RestoreAnalysis Text

The ID of the analysis that you're restoring.

Destructuring the Response

data RestoreAnalysisResponse

See: newRestoreAnalysisResponse smart constructor.

Constructors

RestoreAnalysisResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

analysisId :: Maybe Text

The ID of the analysis that you're restoring.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the analysis that you're restoring.

status :: Int

The HTTP status of the request.

newRestoreAnalysisResponse

Arguments

:: Int
$sel:status:RestoreAnalysisResponse'

-> RestoreAnalysisResponse

Create a value of RestoreAnalysisResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:RestoreAnalysisResponse', restoreAnalysisResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:analysisId:RestoreAnalysis', restoreAnalysisResponse_analysisId - The ID of the analysis that you're restoring.

$sel:arn:RestoreAnalysisResponse', restoreAnalysisResponse_arn - The Amazon Resource Name (ARN) of the analysis that you're restoring.

$sel:status:RestoreAnalysisResponse', restoreAnalysisResponse_status - The HTTP status of the request.

Response Lenses

restoreAnalysisResponse_requestId :: Lens' RestoreAnalysisResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

restoreAnalysisResponse_analysisId :: Lens' RestoreAnalysisResponse (Maybe Text)

The ID of the analysis that you're restoring.

restoreAnalysisResponse_arn :: Lens' RestoreAnalysisResponse (Maybe Text)

The Amazon Resource Name (ARN) of the analysis that you're restoring.

restoreAnalysisResponse_status :: Lens' RestoreAnalysisResponse Int

The HTTP status of the request.