Network.AWS.QuickSight.Types.AccountSettings

Description

Documentation

data AccountSettings

The Amazon QuickSight settings associated with your Amazon Web Services account.

See: newAccountSettings smart constructor.

Constructors

AccountSettings'

Fields

edition :: Maybe Edition

The edition of Amazon QuickSight that you're currently subscribed to: Enterprise edition or Standard edition.

accountName :: Maybe Text

The "account name" you provided for the Amazon QuickSight subscription in your Amazon Web Services account. You create this name when you sign up for Amazon QuickSight. It is unique in all of Amazon Web Services and it appears only when users sign in.

defaultNamespace :: Maybe Text

The default Amazon QuickSight namespace for your Amazon Web Services account.

notificationEmail :: Maybe Text

The main notification email for your Amazon QuickSight subscription.

newAccountSettings :: AccountSettings

Create a value of AccountSettings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:edition:AccountSettings', accountSettings_edition - The edition of Amazon QuickSight that you're currently subscribed to: Enterprise edition or Standard edition.

$sel:accountName:AccountSettings', accountSettings_accountName - The "account name" you provided for the Amazon QuickSight subscription in your Amazon Web Services account. You create this name when you sign up for Amazon QuickSight. It is unique in all of Amazon Web Services and it appears only when users sign in.

$sel:defaultNamespace:AccountSettings', accountSettings_defaultNamespace - The default Amazon QuickSight namespace for your Amazon Web Services account.

$sel:notificationEmail:AccountSettings', accountSettings_notificationEmail - The main notification email for your Amazon QuickSight subscription.

accountSettings_edition :: Lens' AccountSettings (Maybe Edition)

The edition of Amazon QuickSight that you're currently subscribed to: Enterprise edition or Standard edition.

accountSettings_accountName :: Lens' AccountSettings (Maybe Text)

The "account name" you provided for the Amazon QuickSight subscription in your Amazon Web Services account. You create this name when you sign up for Amazon QuickSight. It is unique in all of Amazon Web Services and it appears only when users sign in.

accountSettings_defaultNamespace :: Lens' AccountSettings (Maybe Text)

The default Amazon QuickSight namespace for your Amazon Web Services account.

accountSettings_notificationEmail :: Lens' AccountSettings (Maybe Text)

The main notification email for your Amazon QuickSight subscription.