Network.AWS.QuickSight.Types.Analysis

Description

Documentation

data Analysis

Metadata structure for an analysis in Amazon QuickSight

See: newAnalysis smart constructor.

Constructors

Analysis'

Fields

status :: Maybe ResourceStatus

Status associated with the analysis.

themeArn :: Maybe Text

The ARN of the theme of the analysis.

dataSetArns :: Maybe [Text]

The ARNs of the datasets of the analysis.

sheets :: Maybe [Sheet]

A list of the associated sheets with the unique identifier and name of each sheet.

analysisId :: Maybe Text

The ID of the analysis.

lastUpdatedTime :: Maybe POSIX

The time that the analysis was last updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the analysis.

createdTime :: Maybe POSIX

The time that the analysis was created.

name :: Maybe Text

The descriptive name of the analysis.

errors :: Maybe (NonEmpty AnalysisError)

Errors associated with the analysis.

newAnalysis :: Analysis

Create a value of Analysis with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:Analysis', analysis_status - Status associated with the analysis.

$sel:themeArn:Analysis', analysis_themeArn - The ARN of the theme of the analysis.

$sel:dataSetArns:Analysis', analysis_dataSetArns - The ARNs of the datasets of the analysis.

$sel:sheets:Analysis', analysis_sheets - A list of the associated sheets with the unique identifier and name of each sheet.

$sel:analysisId:Analysis', analysis_analysisId - The ID of the analysis.

$sel:lastUpdatedTime:Analysis', analysis_lastUpdatedTime - The time that the analysis was last updated.

$sel:arn:Analysis', analysis_arn - The Amazon Resource Name (ARN) of the analysis.

$sel:createdTime:Analysis', analysis_createdTime - The time that the analysis was created.

$sel:name:Analysis', analysis_name - The descriptive name of the analysis.

$sel:errors:Analysis', analysis_errors - Errors associated with the analysis.

analysis_status :: Lens' Analysis (Maybe ResourceStatus)

Status associated with the analysis.

analysis_themeArn :: Lens' Analysis (Maybe Text)

The ARN of the theme of the analysis.

analysis_dataSetArns :: Lens' Analysis (Maybe [Text])

The ARNs of the datasets of the analysis.

analysis_sheets :: Lens' Analysis (Maybe [Sheet])

A list of the associated sheets with the unique identifier and name of each sheet.

analysis_analysisId :: Lens' Analysis (Maybe Text)

The ID of the analysis.

analysis_lastUpdatedTime :: Lens' Analysis (Maybe UTCTime)

The time that the analysis was last updated.

analysis_arn :: Lens' Analysis (Maybe Text)

The Amazon Resource Name (ARN) of the analysis.

analysis_createdTime :: Lens' Analysis (Maybe UTCTime)

The time that the analysis was created.

analysis_name :: Lens' Analysis (Maybe Text)

The descriptive name of the analysis.

analysis_errors :: Lens' Analysis (Maybe (NonEmpty AnalysisError))

Errors associated with the analysis.