Network.AWS.QuickSight.Types.DashboardVersion

Description

Documentation

data DashboardVersion

Dashboard version.

See: newDashboardVersion smart constructor.

Constructors

DashboardVersion'

Fields

status :: Maybe ResourceStatus

The HTTP status of the request.

themeArn :: Maybe Text

The ARN of the theme associated with a version of the dashboard.

dataSetArns :: Maybe [Text]

The Amazon Resource Numbers (ARNs) for the datasets that are associated with this version of the dashboard.

sheets :: Maybe [Sheet]

A list of the associated sheets with the unique identifier and name of each sheet.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the resource.

createdTime :: Maybe POSIX

The time that this dashboard version was created.

sourceEntityArn :: Maybe Text

Source entity ARN.

versionNumber :: Maybe Natural

Version number for this version of the dashboard.

errors :: Maybe (NonEmpty DashboardError)

Errors associated with this dashboard version.

description :: Maybe Text

Description.

newDashboardVersion :: DashboardVersion

Create a value of DashboardVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DashboardVersion', dashboardVersion_status - The HTTP status of the request.

$sel:themeArn:DashboardVersion', dashboardVersion_themeArn - The ARN of the theme associated with a version of the dashboard.

$sel:dataSetArns:DashboardVersion', dashboardVersion_dataSetArns - The Amazon Resource Numbers (ARNs) for the datasets that are associated with this version of the dashboard.

$sel:sheets:DashboardVersion', dashboardVersion_sheets - A list of the associated sheets with the unique identifier and name of each sheet.

$sel:arn:DashboardVersion', dashboardVersion_arn - The Amazon Resource Name (ARN) of the resource.

$sel:createdTime:DashboardVersion', dashboardVersion_createdTime - The time that this dashboard version was created.

$sel:sourceEntityArn:DashboardVersion', dashboardVersion_sourceEntityArn - Source entity ARN.

$sel:versionNumber:DashboardVersion', dashboardVersion_versionNumber - Version number for this version of the dashboard.

$sel:errors:DashboardVersion', dashboardVersion_errors - Errors associated with this dashboard version.

$sel:description:DashboardVersion', dashboardVersion_description - Description.

dashboardVersion_status :: Lens' DashboardVersion (Maybe ResourceStatus)

The HTTP status of the request.

dashboardVersion_themeArn :: Lens' DashboardVersion (Maybe Text)

The ARN of the theme associated with a version of the dashboard.

dashboardVersion_dataSetArns :: Lens' DashboardVersion (Maybe [Text])

The Amazon Resource Numbers (ARNs) for the datasets that are associated with this version of the dashboard.

dashboardVersion_sheets :: Lens' DashboardVersion (Maybe [Sheet])

A list of the associated sheets with the unique identifier and name of each sheet.

dashboardVersion_arn :: Lens' DashboardVersion (Maybe Text)

The Amazon Resource Name (ARN) of the resource.

dashboardVersion_createdTime :: Lens' DashboardVersion (Maybe UTCTime)

The time that this dashboard version was created.

dashboardVersion_sourceEntityArn :: Lens' DashboardVersion (Maybe Text)

Source entity ARN.

dashboardVersion_versionNumber :: Lens' DashboardVersion (Maybe Natural)

Version number for this version of the dashboard.

dashboardVersion_errors :: Lens' DashboardVersion (Maybe (NonEmpty DashboardError))

Errors associated with this dashboard version.

dashboardVersion_description :: Lens' DashboardVersion (Maybe Text)

Description.