Network.AWS.QuickSight.Types.DataSource

Description

Documentation

data DataSource

The structure of a data source.

See: newDataSource smart constructor.

Constructors

DataSource'

Fields

status :: Maybe ResourceStatus

The HTTP status of the request.

dataSourceParameters :: Maybe DataSourceParameters

The parameters that Amazon QuickSight uses to connect to your underlying source. This is a variant type structure. For this structure to be valid, only one of the attributes can be non-null.

lastUpdatedTime :: Maybe POSIX

The last time that this data source was updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the data source.

createdTime :: Maybe POSIX

The time that this data source was created.

sslProperties :: Maybe SslProperties

Secure Socket Layer (SSL) properties that apply when Amazon QuickSight connects to your underlying source.

dataSourceId :: Maybe Text

The ID of the data source. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

name :: Maybe Text

A display name for the data source.

alternateDataSourceParameters :: Maybe (NonEmpty DataSourceParameters)

A set of alternate data source parameters that you want to share for the credentials stored with this data source. The credentials are applied in tandem with the data source parameters when you copy a data source by using a create or update request. The API operation compares the DataSourceParameters structure that's in the request with the structures in the AlternateDataSourceParameters allow list. If the structures are an exact match, the request is allowed to use the credentials from this existing data source. If the AlternateDataSourceParameters list is null, the Credentials originally used with this DataSourceParameters are automatically allowed.

vpcConnectionProperties :: Maybe VpcConnectionProperties

The VPC connection information. You need to use this parameter only when you want Amazon QuickSight to use a VPC connection when connecting to your underlying source.

type' :: Maybe DataSourceType

The type of the data source. This type indicates which database engine the data source connects to.

errorInfo :: Maybe DataSourceErrorInfo

Error information from the last update or the creation of the data source.

newDataSource :: DataSource

Create a value of DataSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DataSource', dataSource_status - The HTTP status of the request.

$sel:dataSourceParameters:DataSource', dataSource_dataSourceParameters - The parameters that Amazon QuickSight uses to connect to your underlying source. This is a variant type structure. For this structure to be valid, only one of the attributes can be non-null.

$sel:lastUpdatedTime:DataSource', dataSource_lastUpdatedTime - The last time that this data source was updated.

$sel:arn:DataSource', dataSource_arn - The Amazon Resource Name (ARN) of the data source.

$sel:createdTime:DataSource', dataSource_createdTime - The time that this data source was created.

$sel:sslProperties:DataSource', dataSource_sslProperties - Secure Socket Layer (SSL) properties that apply when Amazon QuickSight connects to your underlying source.

$sel:dataSourceId:DataSource', dataSource_dataSourceId - The ID of the data source. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

$sel:name:DataSource', dataSource_name - A display name for the data source.

$sel:alternateDataSourceParameters:DataSource', dataSource_alternateDataSourceParameters - A set of alternate data source parameters that you want to share for the credentials stored with this data source. The credentials are applied in tandem with the data source parameters when you copy a data source by using a create or update request. The API operation compares the DataSourceParameters structure that's in the request with the structures in the AlternateDataSourceParameters allow list. If the structures are an exact match, the request is allowed to use the credentials from this existing data source. If the AlternateDataSourceParameters list is null, the Credentials originally used with this DataSourceParameters are automatically allowed.

$sel:vpcConnectionProperties:DataSource', dataSource_vpcConnectionProperties - The VPC connection information. You need to use this parameter only when you want Amazon QuickSight to use a VPC connection when connecting to your underlying source.

$sel:type':DataSource', dataSource_type - The type of the data source. This type indicates which database engine the data source connects to.

$sel:errorInfo:DataSource', dataSource_errorInfo - Error information from the last update or the creation of the data source.

dataSource_status :: Lens' DataSource (Maybe ResourceStatus)

The HTTP status of the request.

dataSource_dataSourceParameters :: Lens' DataSource (Maybe DataSourceParameters)

The parameters that Amazon QuickSight uses to connect to your underlying source. This is a variant type structure. For this structure to be valid, only one of the attributes can be non-null.

dataSource_lastUpdatedTime :: Lens' DataSource (Maybe UTCTime)

The last time that this data source was updated.

dataSource_arn :: Lens' DataSource (Maybe Text)

The Amazon Resource Name (ARN) of the data source.

dataSource_createdTime :: Lens' DataSource (Maybe UTCTime)

The time that this data source was created.

dataSource_sslProperties :: Lens' DataSource (Maybe SslProperties)

Secure Socket Layer (SSL) properties that apply when Amazon QuickSight connects to your underlying source.

dataSource_dataSourceId :: Lens' DataSource (Maybe Text)

The ID of the data source. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

dataSource_name :: Lens' DataSource (Maybe Text)

A display name for the data source.

dataSource_alternateDataSourceParameters :: Lens' DataSource (Maybe (NonEmpty DataSourceParameters))

A set of alternate data source parameters that you want to share for the credentials stored with this data source. The credentials are applied in tandem with the data source parameters when you copy a data source by using a create or update request. The API operation compares the DataSourceParameters structure that's in the request with the structures in the AlternateDataSourceParameters allow list. If the structures are an exact match, the request is allowed to use the credentials from this existing data source. If the AlternateDataSourceParameters list is null, the Credentials originally used with this DataSourceParameters are automatically allowed.

dataSource_vpcConnectionProperties :: Lens' DataSource (Maybe VpcConnectionProperties)

The VPC connection information. You need to use this parameter only when you want Amazon QuickSight to use a VPC connection when connecting to your underlying source.

dataSource_type :: Lens' DataSource (Maybe DataSourceType)

The type of the data source. This type indicates which database engine the data source connects to.

dataSource_errorInfo :: Lens' DataSource (Maybe DataSourceErrorInfo)

Error information from the last update or the creation of the data source.