Network.AWS.QuickSight.Types.DataSourceParameters

Description

Documentation

data DataSourceParameters

The parameters that Amazon QuickSight uses to connect to your underlying data source. This is a variant type structure. For this structure to be valid, only one of the attributes can be non-null.

See: newDataSourceParameters smart constructor.

Constructors

DataSourceParameters'

Fields

rdsParameters :: Maybe RdsParameters

The parameters for Amazon RDS.

s3Parameters :: Maybe S3Parameters

The parameters for S3.

athenaParameters :: Maybe AthenaParameters

The parameters for Amazon Athena.

teradataParameters :: Maybe TeradataParameters

The parameters for Teradata.

auroraParameters :: Maybe AuroraParameters

The parameters for Amazon Aurora MySQL.

awsIotAnalyticsParameters :: Maybe AwsIotAnalyticsParameters

The parameters for IoT Analytics.

amazonOpenSearchParameters :: Maybe AmazonOpenSearchParameters

sqlServerParameters :: Maybe SqlServerParameters

The parameters for SQL Server.

redshiftParameters :: Maybe RedshiftParameters

The parameters for Amazon Redshift.

postgreSqlParameters :: Maybe PostgreSqlParameters

The parameters for PostgreSQL.

oracleParameters :: Maybe OracleParameters

The parameters for Oracle.

amazonElasticsearchParameters :: Maybe AmazonElasticsearchParameters

The parameters for OpenSearch.

twitterParameters :: Maybe TwitterParameters

The parameters for Twitter.

mariaDbParameters :: Maybe MariaDbParameters

The parameters for MariaDB.

jiraParameters :: Maybe JiraParameters

The parameters for Jira.

snowflakeParameters :: Maybe SnowflakeParameters

The parameters for Snowflake.

prestoParameters :: Maybe PrestoParameters

The parameters for Presto.

mySqlParameters :: Maybe MySqlParameters

The parameters for MySQL.

auroraPostgreSqlParameters :: Maybe AuroraPostgreSqlParameters

The parameters for Amazon Aurora.

sparkParameters :: Maybe SparkParameters

The parameters for Spark.

serviceNowParameters :: Maybe ServiceNowParameters

The parameters for ServiceNow.

newDataSourceParameters :: DataSourceParameters

Create a value of DataSourceParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rdsParameters:DataSourceParameters', dataSourceParameters_rdsParameters - The parameters for Amazon RDS.

$sel:s3Parameters:DataSourceParameters', dataSourceParameters_s3Parameters - The parameters for S3.

$sel:athenaParameters:DataSourceParameters', dataSourceParameters_athenaParameters - The parameters for Amazon Athena.

$sel:teradataParameters:DataSourceParameters', dataSourceParameters_teradataParameters - The parameters for Teradata.

$sel:auroraParameters:DataSourceParameters', dataSourceParameters_auroraParameters - The parameters for Amazon Aurora MySQL.

$sel:awsIotAnalyticsParameters:DataSourceParameters', dataSourceParameters_awsIotAnalyticsParameters - The parameters for IoT Analytics.

$sel:amazonOpenSearchParameters:DataSourceParameters', dataSourceParameters_amazonOpenSearchParameters - Undocumented member.

$sel:sqlServerParameters:DataSourceParameters', dataSourceParameters_sqlServerParameters - The parameters for SQL Server.

$sel:redshiftParameters:DataSourceParameters', dataSourceParameters_redshiftParameters - The parameters for Amazon Redshift.

$sel:postgreSqlParameters:DataSourceParameters', dataSourceParameters_postgreSqlParameters - The parameters for PostgreSQL.

$sel:oracleParameters:DataSourceParameters', dataSourceParameters_oracleParameters - The parameters for Oracle.

$sel:amazonElasticsearchParameters:DataSourceParameters', dataSourceParameters_amazonElasticsearchParameters - The parameters for OpenSearch.

$sel:twitterParameters:DataSourceParameters', dataSourceParameters_twitterParameters - The parameters for Twitter.

$sel:mariaDbParameters:DataSourceParameters', dataSourceParameters_mariaDbParameters - The parameters for MariaDB.

$sel:jiraParameters:DataSourceParameters', dataSourceParameters_jiraParameters - The parameters for Jira.

$sel:snowflakeParameters:DataSourceParameters', dataSourceParameters_snowflakeParameters - The parameters for Snowflake.

$sel:prestoParameters:DataSourceParameters', dataSourceParameters_prestoParameters - The parameters for Presto.

$sel:mySqlParameters:DataSourceParameters', dataSourceParameters_mySqlParameters - The parameters for MySQL.

$sel:auroraPostgreSqlParameters:DataSourceParameters', dataSourceParameters_auroraPostgreSqlParameters - The parameters for Amazon Aurora.

$sel:sparkParameters:DataSourceParameters', dataSourceParameters_sparkParameters - The parameters for Spark.

$sel:serviceNowParameters:DataSourceParameters', dataSourceParameters_serviceNowParameters - The parameters for ServiceNow.

dataSourceParameters_rdsParameters :: Lens' DataSourceParameters (Maybe RdsParameters)

The parameters for Amazon RDS.

dataSourceParameters_s3Parameters :: Lens' DataSourceParameters (Maybe S3Parameters)

The parameters for S3.

dataSourceParameters_athenaParameters :: Lens' DataSourceParameters (Maybe AthenaParameters)

The parameters for Amazon Athena.

dataSourceParameters_teradataParameters :: Lens' DataSourceParameters (Maybe TeradataParameters)

The parameters for Teradata.

dataSourceParameters_auroraParameters :: Lens' DataSourceParameters (Maybe AuroraParameters)

The parameters for Amazon Aurora MySQL.

dataSourceParameters_awsIotAnalyticsParameters :: Lens' DataSourceParameters (Maybe AwsIotAnalyticsParameters)

The parameters for IoT Analytics.

dataSourceParameters_amazonOpenSearchParameters :: Lens' DataSourceParameters (Maybe AmazonOpenSearchParameters)

Undocumented member.

dataSourceParameters_sqlServerParameters :: Lens' DataSourceParameters (Maybe SqlServerParameters)

The parameters for SQL Server.

dataSourceParameters_redshiftParameters :: Lens' DataSourceParameters (Maybe RedshiftParameters)

The parameters for Amazon Redshift.

dataSourceParameters_postgreSqlParameters :: Lens' DataSourceParameters (Maybe PostgreSqlParameters)

The parameters for PostgreSQL.

dataSourceParameters_oracleParameters :: Lens' DataSourceParameters (Maybe OracleParameters)

The parameters for Oracle.

dataSourceParameters_amazonElasticsearchParameters :: Lens' DataSourceParameters (Maybe AmazonElasticsearchParameters)

The parameters for OpenSearch.

dataSourceParameters_twitterParameters :: Lens' DataSourceParameters (Maybe TwitterParameters)

The parameters for Twitter.

dataSourceParameters_mariaDbParameters :: Lens' DataSourceParameters (Maybe MariaDbParameters)

The parameters for MariaDB.

dataSourceParameters_jiraParameters :: Lens' DataSourceParameters (Maybe JiraParameters)

The parameters for Jira.

dataSourceParameters_snowflakeParameters :: Lens' DataSourceParameters (Maybe SnowflakeParameters)

The parameters for Snowflake.

dataSourceParameters_prestoParameters :: Lens' DataSourceParameters (Maybe PrestoParameters)

The parameters for Presto.

dataSourceParameters_mySqlParameters :: Lens' DataSourceParameters (Maybe MySqlParameters)

The parameters for MySQL.

dataSourceParameters_auroraPostgreSqlParameters :: Lens' DataSourceParameters (Maybe AuroraPostgreSqlParameters)

The parameters for Amazon Aurora.

dataSourceParameters_sparkParameters :: Lens' DataSourceParameters (Maybe SparkParameters)

The parameters for Spark.

dataSourceParameters_serviceNowParameters :: Lens' DataSourceParameters (Maybe ServiceNowParameters)

The parameters for ServiceNow.