Network.AWS.QuickSight.Types.JoinInstruction

Description

Documentation

data JoinInstruction

The instructions associated with a join.

See: newJoinInstruction smart constructor.

Constructors

JoinInstruction'

Fields

rightJoinKeyProperties :: Maybe JoinKeyProperties

Join key properties of the right operand.

leftJoinKeyProperties :: Maybe JoinKeyProperties

Join key properties of the left operand.

leftOperand :: Text

The operand on the left side of a join.

rightOperand :: Text

The operand on the right side of a join.

type' :: JoinType

The type of join that it is.

onClause :: Text

The join instructions provided in the ON clause of a join.

newJoinInstruction

Arguments

:: Text
$sel:leftOperand:JoinInstruction'

-> Text
$sel:rightOperand:JoinInstruction'

-> JoinType
$sel:type':JoinInstruction'

-> Text
$sel:onClause:JoinInstruction'

-> JoinInstruction

Create a value of JoinInstruction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:rightJoinKeyProperties:JoinInstruction', joinInstruction_rightJoinKeyProperties - Join key properties of the right operand.

$sel:leftJoinKeyProperties:JoinInstruction', joinInstruction_leftJoinKeyProperties - Join key properties of the left operand.

$sel:leftOperand:JoinInstruction', joinInstruction_leftOperand - The operand on the left side of a join.

$sel:rightOperand:JoinInstruction', joinInstruction_rightOperand - The operand on the right side of a join.

$sel:type':JoinInstruction', joinInstruction_type - The type of join that it is.

$sel:onClause:JoinInstruction', joinInstruction_onClause - The join instructions provided in the ON clause of a join.

joinInstruction_rightJoinKeyProperties :: Lens' JoinInstruction (Maybe JoinKeyProperties)

Join key properties of the right operand.

joinInstruction_leftJoinKeyProperties :: Lens' JoinInstruction (Maybe JoinKeyProperties)

Join key properties of the left operand.

joinInstruction_leftOperand :: Lens' JoinInstruction Text

The operand on the left side of a join.

joinInstruction_rightOperand :: Lens' JoinInstruction Text

The operand on the right side of a join.

joinInstruction_type :: Lens' JoinInstruction JoinType

The type of join that it is.

joinInstruction_onClause :: Lens' JoinInstruction Text

The join instructions provided in the ON clause of a join.