Network.AWS.QuickSight.Types.LogicalTable

Description

Documentation

data LogicalTable

A logical table is a unit that joins and that data transformations operate on. A logical table has a source, which can be either a physical table or result of a join. When a logical table points to a physical table, the logical table acts as a mutable copy of that physical table through transform operations.

See: newLogicalTable smart constructor.

Constructors

LogicalTable'

Fields

dataTransforms :: Maybe (NonEmpty TransformOperation)

Transform operations that act on this logical table.

alias :: Text

A display name for the logical table.

source :: LogicalTableSource

Source of this logical table.

newLogicalTable

Arguments

:: Text
$sel:alias:LogicalTable'

-> LogicalTableSource
$sel:source:LogicalTable'

-> LogicalTable

Create a value of LogicalTable with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataTransforms:LogicalTable', logicalTable_dataTransforms - Transform operations that act on this logical table.

$sel:alias:LogicalTable', logicalTable_alias - A display name for the logical table.

$sel:source:LogicalTable', logicalTable_source - Source of this logical table.

logicalTable_dataTransforms :: Lens' LogicalTable (Maybe (NonEmpty TransformOperation))

Transform operations that act on this logical table.

logicalTable_alias :: Lens' LogicalTable Text

A display name for the logical table.

logicalTable_source :: Lens' LogicalTable LogicalTableSource

Source of this logical table.