Network.AWS.QuickSight.Types.QueueInfo

Description

Documentation

data QueueInfo

Information about a queued dataset SPICE ingestion.

See: newQueueInfo smart constructor.

Constructors

QueueInfo'

Fields

waitingOnIngestion :: Text

The ID of the queued ingestion.

queuedIngestion :: Text

The ID of the ongoing ingestion. The queued ingestion is waiting for the ongoing ingestion to complete.

newQueueInfo

Arguments

:: Text
$sel:waitingOnIngestion:QueueInfo'

-> Text
$sel:queuedIngestion:QueueInfo'

-> QueueInfo

Create a value of QueueInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:waitingOnIngestion:QueueInfo', queueInfo_waitingOnIngestion - The ID of the queued ingestion.

$sel:queuedIngestion:QueueInfo', queueInfo_queuedIngestion - The ID of the ongoing ingestion. The queued ingestion is waiting for the ongoing ingestion to complete.

queueInfo_waitingOnIngestion :: Lens' QueueInfo Text

The ID of the queued ingestion.

queueInfo_queuedIngestion :: Lens' QueueInfo Text

The ID of the ongoing ingestion. The queued ingestion is waiting for the ongoing ingestion to complete.