Network.AWS.QuickSight.Types.RowLevelPermissionTagRule

Description

Documentation

data RowLevelPermissionTagRule

A set of rules associated with a tag.

See: newRowLevelPermissionTagRule smart constructor.

Constructors

RowLevelPermissionTagRule'

Fields

tagMultiValueDelimiter :: Maybe Text

A string that you want to use to delimit the values when you pass the values at run time. For example, you can delimit the values with a comma.

matchAllValue :: Maybe (Sensitive Text)

A string that you want to use to filter by all the values in a column in the dataset and don’t want to list the values one by one. For example, you can use an asterisk as your match all value.

tagKey :: Text

The unique key for a tag.

columnName :: Text

The column name that a tag key is assigned to.

newRowLevelPermissionTagRule

Arguments

:: Text
$sel:tagKey:RowLevelPermissionTagRule'

-> Text
$sel:columnName:RowLevelPermissionTagRule'

-> RowLevelPermissionTagRule

Create a value of RowLevelPermissionTagRule with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagMultiValueDelimiter:RowLevelPermissionTagRule', rowLevelPermissionTagRule_tagMultiValueDelimiter - A string that you want to use to delimit the values when you pass the values at run time. For example, you can delimit the values with a comma.

$sel:matchAllValue:RowLevelPermissionTagRule', rowLevelPermissionTagRule_matchAllValue - A string that you want to use to filter by all the values in a column in the dataset and don’t want to list the values one by one. For example, you can use an asterisk as your match all value.

$sel:tagKey:RowLevelPermissionTagRule', rowLevelPermissionTagRule_tagKey - The unique key for a tag.

$sel:columnName:RowLevelPermissionTagRule', rowLevelPermissionTagRule_columnName - The column name that a tag key is assigned to.

rowLevelPermissionTagRule_tagMultiValueDelimiter :: Lens' RowLevelPermissionTagRule (Maybe Text)

A string that you want to use to delimit the values when you pass the values at run time. For example, you can delimit the values with a comma.

rowLevelPermissionTagRule_matchAllValue :: Lens' RowLevelPermissionTagRule (Maybe Text)

A string that you want to use to filter by all the values in a column in the dataset and don’t want to list the values one by one. For example, you can use an asterisk as your match all value.

rowLevelPermissionTagRule_tagKey :: Lens' RowLevelPermissionTagRule Text

The unique key for a tag.

rowLevelPermissionTagRule_columnName :: Lens' RowLevelPermissionTagRule Text

The column name that a tag key is assigned to.