Network.AWS.QuickSight.Types.TemplateSummary

Description

Documentation

data TemplateSummary

The template summary.

See: newTemplateSummary smart constructor.

Constructors

TemplateSummary'

Fields

lastUpdatedTime :: Maybe POSIX

The last time that this template was updated.

latestVersionNumber :: Maybe Natural

A structure containing a list of version numbers for the template summary.

arn :: Maybe Text

A summary of a template.

createdTime :: Maybe POSIX

The last time that this template was created.

templateId :: Maybe Text

The ID of the template. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

name :: Maybe Text

A display name for the template.

newTemplateSummary :: TemplateSummary

Create a value of TemplateSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:lastUpdatedTime:TemplateSummary', templateSummary_lastUpdatedTime - The last time that this template was updated.

$sel:latestVersionNumber:TemplateSummary', templateSummary_latestVersionNumber - A structure containing a list of version numbers for the template summary.

$sel:arn:TemplateSummary', templateSummary_arn - A summary of a template.

$sel:createdTime:TemplateSummary', templateSummary_createdTime - The last time that this template was created.

$sel:templateId:TemplateSummary', templateSummary_templateId - The ID of the template. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

$sel:name:TemplateSummary', templateSummary_name - A display name for the template.

templateSummary_lastUpdatedTime :: Lens' TemplateSummary (Maybe UTCTime)

The last time that this template was updated.

templateSummary_latestVersionNumber :: Lens' TemplateSummary (Maybe Natural)

A structure containing a list of version numbers for the template summary.

templateSummary_arn :: Lens' TemplateSummary (Maybe Text)

A summary of a template.

templateSummary_createdTime :: Lens' TemplateSummary (Maybe UTCTime)

The last time that this template was created.

templateSummary_templateId :: Lens' TemplateSummary (Maybe Text)

The ID of the template. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

templateSummary_name :: Lens' TemplateSummary (Maybe Text)

A display name for the template.