Network.AWS.QuickSight.Types.ThemeVersion

Description

Documentation

data ThemeVersion

A version of a theme.

See: newThemeVersion smart constructor.

Constructors

ThemeVersion'

Fields

status :: Maybe ResourceStatus

The status of the theme version.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the resource.

createdTime :: Maybe POSIX

The date and time that this theme version was created.

versionNumber :: Maybe Natural

The version number of the theme.

configuration :: Maybe ThemeConfiguration

The theme configuration, which contains all the theme display properties.

errors :: Maybe (NonEmpty ThemeError)

Errors associated with the theme.

description :: Maybe Text

The description of the theme.

baseThemeId :: Maybe Text

The Amazon QuickSight-defined ID of the theme that a custom theme inherits from. All themes initially inherit from a default Amazon QuickSight theme.

newThemeVersion :: ThemeVersion

Create a value of ThemeVersion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ThemeVersion', themeVersion_status - The status of the theme version.

$sel:arn:ThemeVersion', themeVersion_arn - The Amazon Resource Name (ARN) of the resource.

$sel:createdTime:ThemeVersion', themeVersion_createdTime - The date and time that this theme version was created.

$sel:versionNumber:ThemeVersion', themeVersion_versionNumber - The version number of the theme.

$sel:configuration:ThemeVersion', themeVersion_configuration - The theme configuration, which contains all the theme display properties.

$sel:errors:ThemeVersion', themeVersion_errors - Errors associated with the theme.

$sel:description:ThemeVersion', themeVersion_description - The description of the theme.

$sel:baseThemeId:ThemeVersion', themeVersion_baseThemeId - The Amazon QuickSight-defined ID of the theme that a custom theme inherits from. All themes initially inherit from a default Amazon QuickSight theme.

themeVersion_status :: Lens' ThemeVersion (Maybe ResourceStatus)

The status of the theme version.

themeVersion_arn :: Lens' ThemeVersion (Maybe Text)

The Amazon Resource Name (ARN) of the resource.

themeVersion_createdTime :: Lens' ThemeVersion (Maybe UTCTime)

The date and time that this theme version was created.

themeVersion_versionNumber :: Lens' ThemeVersion (Maybe Natural)

The version number of the theme.

themeVersion_configuration :: Lens' ThemeVersion (Maybe ThemeConfiguration)

The theme configuration, which contains all the theme display properties.

themeVersion_errors :: Lens' ThemeVersion (Maybe (NonEmpty ThemeError))

Errors associated with the theme.

themeVersion_description :: Lens' ThemeVersion (Maybe Text)

The description of the theme.

themeVersion_baseThemeId :: Lens' ThemeVersion (Maybe Text)

The Amazon QuickSight-defined ID of the theme that a custom theme inherits from. All themes initially inherit from a default Amazon QuickSight theme.