Network.AWS.QuickSight.Types.User

Description

Documentation

data User

A registered user of Amazon QuickSight.

See: newUser smart constructor.

Constructors

User'

Fields

email :: Maybe Text

The user's email address.

principalId :: Maybe Text

The principal ID of the user.

arn :: Maybe Text

The Amazon Resource Name (ARN) for the user.

userName :: Maybe Text

The user's user name.

identityType :: Maybe IdentityType

The type of identity authentication used by the user.

externalLoginFederationProviderUrl :: Maybe Text

The URL of the external login provider.

customPermissionsName :: Maybe Text

The custom permissions profile associated with this user.

active :: Maybe Bool

The active status of user. When you create an Amazon QuickSight user that’s not an IAM user or an Active Directory user, that user is inactive until they sign in and provide a password.

role' :: Maybe UserRole

The Amazon QuickSight role for the user. The user role can be one of the following:.

READER: A user who has read-only access to dashboards.

AUTHOR: A user who can create data sources, datasets, analyses, and dashboards.

ADMIN: A user who is an author, who can also manage Amazon Amazon QuickSight settings.

RESTRICTED_READER: This role isn't currently available for use.

RESTRICTED_AUTHOR: This role isn't currently available for use.

externalLoginFederationProviderType :: Maybe Text

The type of supported external login provider that provides identity to let the user federate into Amazon QuickSight with an associated IAMrole. The type can be one of the following.

COGNITO: Amazon Cognito. The provider URL is cognito-identity.amazonaws.com.

CUSTOM_OIDC: Custom OpenID Connect (OIDC) provider.

externalLoginId :: Maybe Text

The identity ID for the user in the external login provider.

newUser :: User

Create a value of User with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:email:User', user_email - The user's email address.

$sel:principalId:User', user_principalId - The principal ID of the user.

$sel:arn:User', user_arn - The Amazon Resource Name (ARN) for the user.

$sel:userName:User', user_userName - The user's user name.

$sel:identityType:User', user_identityType - The type of identity authentication used by the user.

$sel:externalLoginFederationProviderUrl:User', user_externalLoginFederationProviderUrl - The URL of the external login provider.

$sel:customPermissionsName:User', user_customPermissionsName - The custom permissions profile associated with this user.

$sel:active:User', user_active - The active status of user. When you create an Amazon QuickSight user that’s not an IAM user or an Active Directory user, that user is inactive until they sign in and provide a password.

$sel:role':User', user_role - The Amazon QuickSight role for the user. The user role can be one of the following:.

READER: A user who has read-only access to dashboards.

AUTHOR: A user who can create data sources, datasets, analyses, and dashboards.

ADMIN: A user who is an author, who can also manage Amazon Amazon QuickSight settings.

RESTRICTED_READER: This role isn't currently available for use.

RESTRICTED_AUTHOR: This role isn't currently available for use.

$sel:externalLoginFederationProviderType:User', user_externalLoginFederationProviderType - The type of supported external login provider that provides identity to let the user federate into Amazon QuickSight with an associated IAMrole. The type can be one of the following.

COGNITO: Amazon Cognito. The provider URL is cognito-identity.amazonaws.com.

CUSTOM_OIDC: Custom OpenID Connect (OIDC) provider.

$sel:externalLoginId:User', user_externalLoginId - The identity ID for the user in the external login provider.

user_email :: Lens' User (Maybe Text)

The user's email address.

user_principalId :: Lens' User (Maybe Text)

The principal ID of the user.

user_arn :: Lens' User (Maybe Text)

The Amazon Resource Name (ARN) for the user.

user_userName :: Lens' User (Maybe Text)

The user's user name.

user_identityType :: Lens' User (Maybe IdentityType)

The type of identity authentication used by the user.

user_externalLoginFederationProviderUrl :: Lens' User (Maybe Text)

The URL of the external login provider.

user_customPermissionsName :: Lens' User (Maybe Text)

The custom permissions profile associated with this user.

user_active :: Lens' User (Maybe Bool)

The active status of user. When you create an Amazon QuickSight user that’s not an IAM user or an Active Directory user, that user is inactive until they sign in and provide a password.

user_role :: Lens' User (Maybe UserRole)

The Amazon QuickSight role for the user. The user role can be one of the following:.

READER: A user who has read-only access to dashboards.

AUTHOR: A user who can create data sources, datasets, analyses, and dashboards.

ADMIN: A user who is an author, who can also manage Amazon Amazon QuickSight settings.

RESTRICTED_READER: This role isn't currently available for use.

RESTRICTED_AUTHOR: This role isn't currently available for use.

user_externalLoginFederationProviderType :: Lens' User (Maybe Text)

The type of supported external login provider that provides identity to let the user federate into Amazon QuickSight with an associated IAMrole. The type can be one of the following.

COGNITO: Amazon Cognito. The provider URL is cognito-identity.amazonaws.com.

CUSTOM_OIDC: Custom OpenID Connect (OIDC) provider.

user_externalLoginId :: Lens' User (Maybe Text)

The identity ID for the user in the external login provider.