Network.AWS.QuickSight.UpdateDataSetPermissions

Description

Updates the permissions on a dataset.

The permissions resource is arn:aws:quicksight:region:aws-account-id:dataset/data-set-id.

Creating a Request

data UpdateDataSetPermissions

See: newUpdateDataSetPermissions smart constructor.

Constructors

UpdateDataSetPermissions'

Fields

revokePermissions :: Maybe (NonEmpty ResourcePermission)

The resource permissions that you want to revoke from the dataset.

grantPermissions :: Maybe (NonEmpty ResourcePermission)

The resource permissions that you want to grant to the dataset.

awsAccountId :: Text

The Amazon Web Services account ID.

dataSetId :: Text

The ID for the dataset whose permissions you want to update. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

newUpdateDataSetPermissions

Arguments

:: Text
$sel:awsAccountId:UpdateDataSetPermissions'

-> Text
$sel:dataSetId:UpdateDataSetPermissions'

-> UpdateDataSetPermissions

Create a value of UpdateDataSetPermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:revokePermissions:UpdateDataSetPermissions', updateDataSetPermissions_revokePermissions - The resource permissions that you want to revoke from the dataset.

$sel:grantPermissions:UpdateDataSetPermissions', updateDataSetPermissions_grantPermissions - The resource permissions that you want to grant to the dataset.

$sel:awsAccountId:UpdateDataSetPermissions', updateDataSetPermissions_awsAccountId - The Amazon Web Services account ID.

$sel:dataSetId:UpdateDataSetPermissions', updateDataSetPermissions_dataSetId - The ID for the dataset whose permissions you want to update. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

Request Lenses

updateDataSetPermissions_revokePermissions :: Lens' UpdateDataSetPermissions (Maybe (NonEmpty ResourcePermission))

The resource permissions that you want to revoke from the dataset.

updateDataSetPermissions_grantPermissions :: Lens' UpdateDataSetPermissions (Maybe (NonEmpty ResourcePermission))

The resource permissions that you want to grant to the dataset.

updateDataSetPermissions_awsAccountId :: Lens' UpdateDataSetPermissions Text

The Amazon Web Services account ID.

updateDataSetPermissions_dataSetId :: Lens' UpdateDataSetPermissions Text

The ID for the dataset whose permissions you want to update. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

Destructuring the Response

data UpdateDataSetPermissionsResponse

See: newUpdateDataSetPermissionsResponse smart constructor.

Constructors

UpdateDataSetPermissionsResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

dataSetArn :: Maybe Text

The Amazon Resource Name (ARN) of the dataset.

dataSetId :: Maybe Text

The ID for the dataset whose permissions you want to update. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

status :: Int

The HTTP status of the request.

newUpdateDataSetPermissionsResponse

Arguments

:: Int
$sel:status:UpdateDataSetPermissionsResponse'

-> UpdateDataSetPermissionsResponse

Create a value of UpdateDataSetPermissionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:UpdateDataSetPermissionsResponse', updateDataSetPermissionsResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:dataSetArn:UpdateDataSetPermissionsResponse', updateDataSetPermissionsResponse_dataSetArn - The Amazon Resource Name (ARN) of the dataset.

$sel:dataSetId:UpdateDataSetPermissions', updateDataSetPermissionsResponse_dataSetId - The ID for the dataset whose permissions you want to update. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

$sel:status:UpdateDataSetPermissionsResponse', updateDataSetPermissionsResponse_status - The HTTP status of the request.

Response Lenses

updateDataSetPermissionsResponse_requestId :: Lens' UpdateDataSetPermissionsResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

updateDataSetPermissionsResponse_dataSetArn :: Lens' UpdateDataSetPermissionsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the dataset.

updateDataSetPermissionsResponse_dataSetId :: Lens' UpdateDataSetPermissionsResponse (Maybe Text)

The ID for the dataset whose permissions you want to update. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

updateDataSetPermissionsResponse_status :: Lens' UpdateDataSetPermissionsResponse Int

The HTTP status of the request.