Network.AWS.QuickSight.UpdateDataSource

Description

Updates a data source.

Creating a Request

data UpdateDataSource

See: newUpdateDataSource smart constructor.

Constructors

UpdateDataSource'

Fields

dataSourceParameters :: Maybe DataSourceParameters

The parameters that Amazon QuickSight uses to connect to your underlying source.

sslProperties :: Maybe SslProperties

Secure Socket Layer (SSL) properties that apply when Amazon QuickSight connects to your underlying source.

credentials :: Maybe (Sensitive DataSourceCredentials)

The credentials that Amazon QuickSight that uses to connect to your underlying source. Currently, only credentials based on user name and password are supported.

vpcConnectionProperties :: Maybe VpcConnectionProperties

Use this parameter only when you want Amazon QuickSight to use a VPC connection when connecting to your underlying source.

awsAccountId :: Text

The Amazon Web Services account ID.

dataSourceId :: Text

The ID of the data source. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

name :: Text

A display name for the data source.

newUpdateDataSource

Arguments

:: Text
$sel:awsAccountId:UpdateDataSource'

-> Text
$sel:dataSourceId:UpdateDataSource'

-> Text
$sel:name:UpdateDataSource'

-> UpdateDataSource

Create a value of UpdateDataSource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataSourceParameters:UpdateDataSource', updateDataSource_dataSourceParameters - The parameters that Amazon QuickSight uses to connect to your underlying source.

$sel:sslProperties:UpdateDataSource', updateDataSource_sslProperties - Secure Socket Layer (SSL) properties that apply when Amazon QuickSight connects to your underlying source.

$sel:credentials:UpdateDataSource', updateDataSource_credentials - The credentials that Amazon QuickSight that uses to connect to your underlying source. Currently, only credentials based on user name and password are supported.

$sel:vpcConnectionProperties:UpdateDataSource', updateDataSource_vpcConnectionProperties - Use this parameter only when you want Amazon QuickSight to use a VPC connection when connecting to your underlying source.

$sel:awsAccountId:UpdateDataSource', updateDataSource_awsAccountId - The Amazon Web Services account ID.

$sel:dataSourceId:UpdateDataSource', updateDataSource_dataSourceId - The ID of the data source. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

$sel:name:UpdateDataSource', updateDataSource_name - A display name for the data source.

Request Lenses

updateDataSource_dataSourceParameters :: Lens' UpdateDataSource (Maybe DataSourceParameters)

The parameters that Amazon QuickSight uses to connect to your underlying source.

updateDataSource_sslProperties :: Lens' UpdateDataSource (Maybe SslProperties)

Secure Socket Layer (SSL) properties that apply when Amazon QuickSight connects to your underlying source.

updateDataSource_credentials :: Lens' UpdateDataSource (Maybe DataSourceCredentials)

The credentials that Amazon QuickSight that uses to connect to your underlying source. Currently, only credentials based on user name and password are supported.

updateDataSource_vpcConnectionProperties :: Lens' UpdateDataSource (Maybe VpcConnectionProperties)

Use this parameter only when you want Amazon QuickSight to use a VPC connection when connecting to your underlying source.

updateDataSource_awsAccountId :: Lens' UpdateDataSource Text

The Amazon Web Services account ID.

updateDataSource_dataSourceId :: Lens' UpdateDataSource Text

The ID of the data source. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

updateDataSource_name :: Lens' UpdateDataSource Text

A display name for the data source.

Destructuring the Response

data UpdateDataSourceResponse

See: newUpdateDataSourceResponse smart constructor.

Constructors

UpdateDataSourceResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the data source.

dataSourceId :: Maybe Text

The ID of the data source. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

updateStatus :: Maybe ResourceStatus

The update status of the data source's last update.

status :: Int

The HTTP status of the request.

newUpdateDataSourceResponse

Arguments

:: Int
$sel:status:UpdateDataSourceResponse'

-> UpdateDataSourceResponse

Create a value of UpdateDataSourceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:UpdateDataSourceResponse', updateDataSourceResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:arn:UpdateDataSourceResponse', updateDataSourceResponse_arn - The Amazon Resource Name (ARN) of the data source.

$sel:dataSourceId:UpdateDataSource', updateDataSourceResponse_dataSourceId - The ID of the data source. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

$sel:updateStatus:UpdateDataSourceResponse', updateDataSourceResponse_updateStatus - The update status of the data source's last update.

$sel:status:UpdateDataSourceResponse', updateDataSourceResponse_status - The HTTP status of the request.

Response Lenses

updateDataSourceResponse_requestId :: Lens' UpdateDataSourceResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

updateDataSourceResponse_arn :: Lens' UpdateDataSourceResponse (Maybe Text)

The Amazon Resource Name (ARN) of the data source.

updateDataSourceResponse_dataSourceId :: Lens' UpdateDataSourceResponse (Maybe Text)

The ID of the data source. This ID is unique per Amazon Web Services Region for each Amazon Web Services account.

updateDataSourceResponse_updateStatus :: Lens' UpdateDataSourceResponse (Maybe ResourceStatus)

The update status of the data source's last update.

updateDataSourceResponse_status :: Lens' UpdateDataSourceResponse Int

The HTTP status of the request.