Network.AWS.QuickSight.UpdateTemplatePermissions

Description

Updates the resource permissions for a template.

Creating a Request

data UpdateTemplatePermissions

See: newUpdateTemplatePermissions smart constructor.

Constructors

UpdateTemplatePermissions'

Fields

revokePermissions :: Maybe [ResourcePermission]

A list of resource permissions to be revoked from the template.

grantPermissions :: Maybe [ResourcePermission]

A list of resource permissions to be granted on the template.

awsAccountId :: Text

The ID of the Amazon Web Services account that contains the template.

templateId :: Text

The ID for the template.

newUpdateTemplatePermissions

Arguments

:: Text
$sel:awsAccountId:UpdateTemplatePermissions'

-> Text
$sel:templateId:UpdateTemplatePermissions'

-> UpdateTemplatePermissions

Create a value of UpdateTemplatePermissions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:revokePermissions:UpdateTemplatePermissions', updateTemplatePermissions_revokePermissions - A list of resource permissions to be revoked from the template.

$sel:grantPermissions:UpdateTemplatePermissions', updateTemplatePermissions_grantPermissions - A list of resource permissions to be granted on the template.

$sel:awsAccountId:UpdateTemplatePermissions', updateTemplatePermissions_awsAccountId - The ID of the Amazon Web Services account that contains the template.

$sel:templateId:UpdateTemplatePermissions', updateTemplatePermissions_templateId - The ID for the template.

Request Lenses

updateTemplatePermissions_revokePermissions :: Lens' UpdateTemplatePermissions (Maybe [ResourcePermission])

A list of resource permissions to be revoked from the template.

updateTemplatePermissions_grantPermissions :: Lens' UpdateTemplatePermissions (Maybe [ResourcePermission])

A list of resource permissions to be granted on the template.

updateTemplatePermissions_awsAccountId :: Lens' UpdateTemplatePermissions Text

The ID of the Amazon Web Services account that contains the template.

updateTemplatePermissions_templateId :: Lens' UpdateTemplatePermissions Text

The ID for the template.

Destructuring the Response

data UpdateTemplatePermissionsResponse

See: newUpdateTemplatePermissionsResponse smart constructor.

Constructors

UpdateTemplatePermissionsResponse'

Fields

requestId :: Maybe Text

The Amazon Web Services request ID for this operation.

templateId :: Maybe Text

The ID for the template.

templateArn :: Maybe Text

The Amazon Resource Name (ARN) of the template.

permissions :: Maybe (NonEmpty ResourcePermission)

A list of resource permissions to be set on the template.

status :: Int

The HTTP status of the request.

newUpdateTemplatePermissionsResponse

Arguments

:: Int
$sel:status:UpdateTemplatePermissionsResponse'

-> UpdateTemplatePermissionsResponse

Create a value of UpdateTemplatePermissionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:requestId:UpdateTemplatePermissionsResponse', updateTemplatePermissionsResponse_requestId - The Amazon Web Services request ID for this operation.

$sel:templateId:UpdateTemplatePermissions', updateTemplatePermissionsResponse_templateId - The ID for the template.

$sel:templateArn:UpdateTemplatePermissionsResponse', updateTemplatePermissionsResponse_templateArn - The Amazon Resource Name (ARN) of the template.

$sel:permissions:UpdateTemplatePermissionsResponse', updateTemplatePermissionsResponse_permissions - A list of resource permissions to be set on the template.

$sel:status:UpdateTemplatePermissionsResponse', updateTemplatePermissionsResponse_status - The HTTP status of the request.

Response Lenses

updateTemplatePermissionsResponse_requestId :: Lens' UpdateTemplatePermissionsResponse (Maybe Text)

The Amazon Web Services request ID for this operation.

updateTemplatePermissionsResponse_templateId :: Lens' UpdateTemplatePermissionsResponse (Maybe Text)

The ID for the template.

updateTemplatePermissionsResponse_templateArn :: Lens' UpdateTemplatePermissionsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the template.

updateTemplatePermissionsResponse_permissions :: Lens' UpdateTemplatePermissionsResponse (Maybe (NonEmpty ResourcePermission))

A list of resource permissions to be set on the template.

updateTemplatePermissionsResponse_status :: Lens' UpdateTemplatePermissionsResponse Int

The HTTP status of the request.