Network.AWS.RAM.DeleteResourceShare

Description

Deletes the specified resource share.

Creating a Request

data DeleteResourceShare

See: newDeleteResourceShare smart constructor.

Constructors

DeleteResourceShare'

Fields

clientToken :: Maybe Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

resourceShareArn :: Text

The Amazon Resource Name (ARN) of the resource share.

newDeleteResourceShare

Arguments

:: Text
$sel:resourceShareArn:DeleteResourceShare'

-> DeleteResourceShare

Create a value of DeleteResourceShare with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:DeleteResourceShare', deleteResourceShare_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:resourceShareArn:DeleteResourceShare', deleteResourceShare_resourceShareArn - The Amazon Resource Name (ARN) of the resource share.

Request Lenses

deleteResourceShare_clientToken :: Lens' DeleteResourceShare (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

deleteResourceShare_resourceShareArn :: Lens' DeleteResourceShare Text

The Amazon Resource Name (ARN) of the resource share.

Destructuring the Response

data DeleteResourceShareResponse

See: newDeleteResourceShareResponse smart constructor.

Constructors

DeleteResourceShareResponse'

Fields

clientToken :: Maybe Text

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

returnValue :: Maybe Bool

Indicates whether the request succeeded.

httpStatus :: Int

The response's http status code.

newDeleteResourceShareResponse

Arguments

:: Int
$sel:httpStatus:DeleteResourceShareResponse'

-> DeleteResourceShareResponse

Create a value of DeleteResourceShareResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clientToken:DeleteResourceShare', deleteResourceShareResponse_clientToken - A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

$sel:returnValue:DeleteResourceShareResponse', deleteResourceShareResponse_returnValue - Indicates whether the request succeeded.

$sel:httpStatus:DeleteResourceShareResponse', deleteResourceShareResponse_httpStatus - The response's http status code.

Response Lenses

deleteResourceShareResponse_clientToken :: Lens' DeleteResourceShareResponse (Maybe Text)

A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

deleteResourceShareResponse_returnValue :: Lens' DeleteResourceShareResponse (Maybe Bool)

Indicates whether the request succeeded.

deleteResourceShareResponse_httpStatus :: Lens' DeleteResourceShareResponse Int

The response's http status code.