Network.AWS.RAM.GetResourcePolicies

Description

Gets the policies for the specified resources that you own and have shared.

This operation returns paginated results.

Creating a Request

data GetResourcePolicies

See: newGetResourcePolicies smart constructor.

Constructors

GetResourcePolicies'

Fields

nextToken :: Maybe Text

The token for the next page of results.

principal :: Maybe Text

The principal.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

resourceArns :: [Text]

The Amazon Resource Names (ARNs) of the resources.

newGetResourcePolicies :: GetResourcePolicies

Create a value of GetResourcePolicies with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetResourcePolicies', getResourcePolicies_nextToken - The token for the next page of results.

$sel:principal:GetResourcePolicies', getResourcePolicies_principal - The principal.

$sel:maxResults:GetResourcePolicies', getResourcePolicies_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

$sel:resourceArns:GetResourcePolicies', getResourcePolicies_resourceArns - The Amazon Resource Names (ARNs) of the resources.

Request Lenses

getResourcePolicies_nextToken :: Lens' GetResourcePolicies (Maybe Text)

The token for the next page of results.

getResourcePolicies_principal :: Lens' GetResourcePolicies (Maybe Text)

The principal.

getResourcePolicies_maxResults :: Lens' GetResourcePolicies (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

getResourcePolicies_resourceArns :: Lens' GetResourcePolicies [Text]

The Amazon Resource Names (ARNs) of the resources.

Destructuring the Response

data GetResourcePoliciesResponse

See: newGetResourcePoliciesResponse smart constructor.

Constructors

GetResourcePoliciesResponse'

Fields

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

policies :: Maybe [Text]

A key policy document, in JSON format.

httpStatus :: Int

The response's http status code.

newGetResourcePoliciesResponse

Arguments

:: Int
$sel:httpStatus:GetResourcePoliciesResponse'

-> GetResourcePoliciesResponse

Create a value of GetResourcePoliciesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nextToken:GetResourcePolicies', getResourcePoliciesResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:policies:GetResourcePoliciesResponse', getResourcePoliciesResponse_policies - A key policy document, in JSON format.

$sel:httpStatus:GetResourcePoliciesResponse', getResourcePoliciesResponse_httpStatus - The response's http status code.

Response Lenses

getResourcePoliciesResponse_nextToken :: Lens' GetResourcePoliciesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

getResourcePoliciesResponse_policies :: Lens' GetResourcePoliciesResponse (Maybe [Text])

A key policy document, in JSON format.

getResourcePoliciesResponse_httpStatus :: Lens' GetResourcePoliciesResponse Int

The response's http status code.