Network.AWS.RAM.ListPrincipals

Description

Lists the principals that you have shared resources with or that have shared resources with you.

This operation returns paginated results.

Creating a Request

data ListPrincipals

See: newListPrincipals smart constructor.

Constructors

ListPrincipals'

Fields

resourceType :: Maybe Text

The resource type.

Valid values: acm-pca:CertificateAuthority | appmesh:Mesh | codebuild:Project | codebuild:ReportGroup | ec2:CapacityReservation | ec2:DedicatedHost | ec2:LocalGatewayRouteTable | ec2:PrefixList | ec2:Subnet | ec2:TrafficMirrorTarget | ec2:TransitGateway | imagebuilder:Component | imagebuilder:Image | imagebuilder:ImageRecipe | imagebuilder:ContainerRecipe | glue:Catalog | glue:Database | glue:Table | license-manager:LicenseConfiguration I network-firewall:FirewallPolicy | network-firewall:StatefulRuleGroup | network-firewall:StatelessRuleGroup | outposts:Outpost | resource-groups:Group | rds:Cluster | route53resolver:FirewallRuleGroup |route53resolver:ResolverQueryLogConfig | route53resolver:ResolverRule | s3-outposts:Outpost | ssm-contacts:Contact | ssm-incidents:ResponsePlan

principals :: Maybe [Text]

The principals.

nextToken :: Maybe Text

The token for the next page of results.

resourceArn :: Maybe Text

The Amazon Resource Name (ARN) of the resource.

maxResults :: Maybe Natural

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

resourceShareArns :: Maybe [Text]

The Amazon Resource Names (ARN) of the resource shares.

resourceOwner :: ResourceOwner

The type of owner.

newListPrincipals

Arguments

:: ResourceOwner
$sel:resourceOwner:ListPrincipals'

-> ListPrincipals

Create a value of ListPrincipals with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:ListPrincipals', listPrincipals_resourceType - The resource type.

Valid values: acm-pca:CertificateAuthority | appmesh:Mesh | codebuild:Project | codebuild:ReportGroup | ec2:CapacityReservation | ec2:DedicatedHost | ec2:LocalGatewayRouteTable | ec2:PrefixList | ec2:Subnet | ec2:TrafficMirrorTarget | ec2:TransitGateway | imagebuilder:Component | imagebuilder:Image | imagebuilder:ImageRecipe | imagebuilder:ContainerRecipe | glue:Catalog | glue:Database | glue:Table | license-manager:LicenseConfiguration I network-firewall:FirewallPolicy | network-firewall:StatefulRuleGroup | network-firewall:StatelessRuleGroup | outposts:Outpost | resource-groups:Group | rds:Cluster | route53resolver:FirewallRuleGroup |route53resolver:ResolverQueryLogConfig | route53resolver:ResolverRule | s3-outposts:Outpost | ssm-contacts:Contact | ssm-incidents:ResponsePlan

$sel:principals:ListPrincipals', listPrincipals_principals - The principals.

$sel:nextToken:ListPrincipals', listPrincipals_nextToken - The token for the next page of results.

$sel:resourceArn:ListPrincipals', listPrincipals_resourceArn - The Amazon Resource Name (ARN) of the resource.

$sel:maxResults:ListPrincipals', listPrincipals_maxResults - The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

$sel:resourceShareArns:ListPrincipals', listPrincipals_resourceShareArns - The Amazon Resource Names (ARN) of the resource shares.

$sel:resourceOwner:ListPrincipals', listPrincipals_resourceOwner - The type of owner.

Request Lenses

listPrincipals_resourceType :: Lens' ListPrincipals (Maybe Text)

The resource type.

Valid values: acm-pca:CertificateAuthority | appmesh:Mesh | codebuild:Project | codebuild:ReportGroup | ec2:CapacityReservation | ec2:DedicatedHost | ec2:LocalGatewayRouteTable | ec2:PrefixList | ec2:Subnet | ec2:TrafficMirrorTarget | ec2:TransitGateway | imagebuilder:Component | imagebuilder:Image | imagebuilder:ImageRecipe | imagebuilder:ContainerRecipe | glue:Catalog | glue:Database | glue:Table | license-manager:LicenseConfiguration I network-firewall:FirewallPolicy | network-firewall:StatefulRuleGroup | network-firewall:StatelessRuleGroup | outposts:Outpost | resource-groups:Group | rds:Cluster | route53resolver:FirewallRuleGroup |route53resolver:ResolverQueryLogConfig | route53resolver:ResolverRule | s3-outposts:Outpost | ssm-contacts:Contact | ssm-incidents:ResponsePlan

listPrincipals_principals :: Lens' ListPrincipals (Maybe [Text])

The principals.

listPrincipals_nextToken :: Lens' ListPrincipals (Maybe Text)

The token for the next page of results.

listPrincipals_resourceArn :: Lens' ListPrincipals (Maybe Text)

The Amazon Resource Name (ARN) of the resource.

listPrincipals_maxResults :: Lens' ListPrincipals (Maybe Natural)

The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

listPrincipals_resourceShareArns :: Lens' ListPrincipals (Maybe [Text])

The Amazon Resource Names (ARN) of the resource shares.

listPrincipals_resourceOwner :: Lens' ListPrincipals ResourceOwner

The type of owner.

Destructuring the Response

data ListPrincipalsResponse

See: newListPrincipalsResponse smart constructor.

Constructors

ListPrincipalsResponse'

Fields

principals :: Maybe [Principal]

The principals.

nextToken :: Maybe Text

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

httpStatus :: Int

The response's http status code.

newListPrincipalsResponse

Arguments

:: Int
$sel:httpStatus:ListPrincipalsResponse'

-> ListPrincipalsResponse

Create a value of ListPrincipalsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:principals:ListPrincipals', listPrincipalsResponse_principals - The principals.

$sel:nextToken:ListPrincipals', listPrincipalsResponse_nextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

$sel:httpStatus:ListPrincipalsResponse', listPrincipalsResponse_httpStatus - The response's http status code.

Response Lenses

listPrincipalsResponse_principals :: Lens' ListPrincipalsResponse (Maybe [Principal])

The principals.

listPrincipalsResponse_nextToken :: Lens' ListPrincipalsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

listPrincipalsResponse_httpStatus :: Lens' ListPrincipalsResponse Int

The response's http status code.