Network.AWS.RAM.Types

Description

Service Configuration

defaultService :: Service

API version 2018-01-04 of the Amazon Resource Access Manager SDK configuration.

Errors

_InvalidClientTokenException :: AsError a => Getting (First ServiceError) a ServiceError

A client token is not valid.

_TagPolicyViolationException :: AsError a => Getting (First ServiceError) a ServiceError

The specified tag is a reserved word and cannot be used.

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

A parameter is not valid.

_InvalidResourceTypeException :: AsError a => Getting (First ServiceError) a ServiceError

The specified resource type is not valid.

_ResourceShareInvitationArnNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The Amazon Resource Name (ARN) for an invitation was not found.

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The specified value for NextToken is not valid.

_MalformedArnException :: AsError a => Getting (First ServiceError) a ServiceError

The format of an Amazon Resource Name (ARN) is not valid.

_ServerInternalException :: AsError a => Getting (First ServiceError) a ServiceError

The service could not respond to the request due to an internal problem.

_UnknownResourceException :: AsError a => Getting (First ServiceError) a ServiceError

A specified resource was not found.

_OperationNotPermittedException :: AsError a => Getting (First ServiceError) a ServiceError

The requested operation is not permitted.

_InvalidMaxResultsException :: AsError a => Getting (First ServiceError) a ServiceError

The specified value for MaxResults is not valid.

_ServiceUnavailableException :: AsError a => Getting (First ServiceError) a ServiceError

The service is not available.

_ResourceArnNotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

An Amazon Resource Name (ARN) was not found.

_IdempotentParameterMismatchException :: AsError a => Getting (First ServiceError) a ServiceError

A client token input parameter was reused with an operation, but at least one of the other input parameters is different from the previous call to the operation.

_MissingRequiredParameterException :: AsError a => Getting (First ServiceError) a ServiceError

A required input parameter is missing.

_TagLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The requested tags exceed the limit for your account.

_InvalidStateTransitionException :: AsError a => Getting (First ServiceError) a ServiceError

The requested state transition is not valid.

_ResourceShareInvitationAlreadyAcceptedException :: AsError a => Getting (First ServiceError) a ServiceError

The invitation was already accepted.

_ResourceShareLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The requested resource share exceeds the limit for your account.

_ResourceShareInvitationExpiredException :: AsError a => Getting (First ServiceError) a ServiceError

The invitation is expired.

_ResourceShareInvitationAlreadyRejectedException :: AsError a => Getting (First ServiceError) a ServiceError

The invitation was already rejected.

ResourceOwner

newtype ResourceOwner

Constructors

ResourceOwner'

Fields

fromResourceOwner :: Text

Bundled Patterns

pattern ResourceOwner_OTHER_ACCOUNTS :: ResourceOwner

pattern ResourceOwner_SELF :: ResourceOwner

ResourceShareAssociationStatus

newtype ResourceShareAssociationStatus

Constructors

ResourceShareAssociationStatus'

Fields

fromResourceShareAssociationStatus :: Text

Bundled Patterns

pattern ResourceShareAssociationStatus_ASSOCIATED :: ResourceShareAssociationStatus

pattern ResourceShareAssociationStatus_ASSOCIATING :: ResourceShareAssociationStatus

pattern ResourceShareAssociationStatus_DISASSOCIATED :: ResourceShareAssociationStatus

pattern ResourceShareAssociationStatus_DISASSOCIATING :: ResourceShareAssociationStatus

pattern ResourceShareAssociationStatus_FAILED :: ResourceShareAssociationStatus

ResourceShareAssociationType

newtype ResourceShareAssociationType

Constructors

ResourceShareAssociationType'

Fields

fromResourceShareAssociationType :: Text

Bundled Patterns

pattern ResourceShareAssociationType_PRINCIPAL :: ResourceShareAssociationType

pattern ResourceShareAssociationType_RESOURCE :: ResourceShareAssociationType

ResourceShareFeatureSet

newtype ResourceShareFeatureSet

Constructors

ResourceShareFeatureSet'

Fields

fromResourceShareFeatureSet :: Text

Bundled Patterns

pattern ResourceShareFeatureSet_CREATED_FROM_POLICY :: ResourceShareFeatureSet

pattern ResourceShareFeatureSet_PROMOTING_TO_STANDARD :: ResourceShareFeatureSet

pattern ResourceShareFeatureSet_STANDARD :: ResourceShareFeatureSet

ResourceShareInvitationStatus

newtype ResourceShareInvitationStatus

Constructors

ResourceShareInvitationStatus'

Fields

fromResourceShareInvitationStatus :: Text

Bundled Patterns

pattern ResourceShareInvitationStatus_ACCEPTED :: ResourceShareInvitationStatus

pattern ResourceShareInvitationStatus_EXPIRED :: ResourceShareInvitationStatus

pattern ResourceShareInvitationStatus_PENDING :: ResourceShareInvitationStatus

pattern ResourceShareInvitationStatus_REJECTED :: ResourceShareInvitationStatus

ResourceShareStatus

newtype ResourceShareStatus

Constructors

ResourceShareStatus'

Fields

fromResourceShareStatus :: Text

Bundled Patterns

pattern ResourceShareStatus_ACTIVE :: ResourceShareStatus

pattern ResourceShareStatus_DELETED :: ResourceShareStatus

pattern ResourceShareStatus_DELETING :: ResourceShareStatus

pattern ResourceShareStatus_FAILED :: ResourceShareStatus

pattern ResourceShareStatus_PENDING :: ResourceShareStatus

ResourceStatus

newtype ResourceStatus

Constructors

ResourceStatus'

Fields

fromResourceStatus :: Text

Bundled Patterns

pattern ResourceStatus_AVAILABLE :: ResourceStatus

pattern ResourceStatus_LIMIT_EXCEEDED :: ResourceStatus

pattern ResourceStatus_PENDING :: ResourceStatus

pattern ResourceStatus_UNAVAILABLE :: ResourceStatus

pattern ResourceStatus_ZONAL_RESOURCE_INACCESSIBLE :: ResourceStatus

Principal

data Principal

Describes a principal for use with Resource Access Manager.

See: newPrincipal smart constructor.

Constructors

Principal'

Fields

creationTime :: Maybe POSIX

The time when the principal was associated with the resource share.

resourceShareArn :: Maybe Text

The Amazon Resource Name (ARN) of the resource share.

lastUpdatedTime :: Maybe POSIX

The time when the association was last updated.

external :: Maybe Bool

Indicates whether the principal belongs to the same organization in Organizations as the Amazon Web Services account that owns the resource share.

id :: Maybe Text

The ID of the principal.

newPrincipal :: Principal

Create a value of Principal with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Principal', principal_creationTime - The time when the principal was associated with the resource share.

$sel:resourceShareArn:Principal', principal_resourceShareArn - The Amazon Resource Name (ARN) of the resource share.

$sel:lastUpdatedTime:Principal', principal_lastUpdatedTime - The time when the association was last updated.

$sel:external:Principal', principal_external - Indicates whether the principal belongs to the same organization in Organizations as the Amazon Web Services account that owns the resource share.

$sel:id:Principal', principal_id - The ID of the principal.

principal_creationTime :: Lens' Principal (Maybe UTCTime)

The time when the principal was associated with the resource share.

principal_resourceShareArn :: Lens' Principal (Maybe Text)

The Amazon Resource Name (ARN) of the resource share.

principal_lastUpdatedTime :: Lens' Principal (Maybe UTCTime)

The time when the association was last updated.

principal_external :: Lens' Principal (Maybe Bool)

Indicates whether the principal belongs to the same organization in Organizations as the Amazon Web Services account that owns the resource share.

principal_id :: Lens' Principal (Maybe Text)

The ID of the principal.

Resource

data Resource

Describes a resource associated with a resource share.

See: newResource smart constructor.

Constructors

Resource'

Fields

creationTime :: Maybe POSIX

The time when the resource was associated with the resource share.

status :: Maybe ResourceStatus

The status of the resource.

resourceShareArn :: Maybe Text

The Amazon Resource Name (ARN) of the resource share.

lastUpdatedTime :: Maybe POSIX

The time when the association was last updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the resource.

resourceGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of the resource group. This value is returned only if the resource is a resource group.

statusMessage :: Maybe Text

A message about the status of the resource.

type' :: Maybe Text

The resource type.

newResource :: Resource

Create a value of Resource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Resource', resource_creationTime - The time when the resource was associated with the resource share.

$sel:status:Resource', resource_status - The status of the resource.

$sel:resourceShareArn:Resource', resource_resourceShareArn - The Amazon Resource Name (ARN) of the resource share.

$sel:lastUpdatedTime:Resource', resource_lastUpdatedTime - The time when the association was last updated.

$sel:arn:Resource', resource_arn - The Amazon Resource Name (ARN) of the resource.

$sel:resourceGroupArn:Resource', resource_resourceGroupArn - The Amazon Resource Name (ARN) of the resource group. This value is returned only if the resource is a resource group.

$sel:statusMessage:Resource', resource_statusMessage - A message about the status of the resource.

$sel:type':Resource', resource_type - The resource type.

resource_creationTime :: Lens' Resource (Maybe UTCTime)

The time when the resource was associated with the resource share.

resource_status :: Lens' Resource (Maybe ResourceStatus)

The status of the resource.

resource_resourceShareArn :: Lens' Resource (Maybe Text)

The Amazon Resource Name (ARN) of the resource share.

resource_lastUpdatedTime :: Lens' Resource (Maybe UTCTime)

The time when the association was last updated.

resource_arn :: Lens' Resource (Maybe Text)

The Amazon Resource Name (ARN) of the resource.

resource_resourceGroupArn :: Lens' Resource (Maybe Text)

The Amazon Resource Name (ARN) of the resource group. This value is returned only if the resource is a resource group.

resource_statusMessage :: Lens' Resource (Maybe Text)

A message about the status of the resource.

resource_type :: Lens' Resource (Maybe Text)

The resource type.

ResourceShare

data ResourceShare

Describes a resource share.

See: newResourceShare smart constructor.

Constructors

ResourceShare'

Fields

creationTime :: Maybe POSIX

The time when the resource share was created.

status :: Maybe ResourceShareStatus

The status of the resource share.

resourceShareArn :: Maybe Text

The Amazon Resource Name (ARN) of the resource share.

owningAccountId :: Maybe Text

The ID of the Amazon Web Services account that owns the resource share.

lastUpdatedTime :: Maybe POSIX

The time when the resource share was last updated.

allowExternalPrincipals :: Maybe Bool

Indicates whether principals outside your organization in Organizations can be associated with a resource share.

name :: Maybe Text

The name of the resource share.

statusMessage :: Maybe Text

A message about the status of the resource share.

featureSet :: Maybe ResourceShareFeatureSet

Indicates how the resource share was created. Possible values include:

CREATED_FROM_POLICY - Indicates that the resource share was created from an Amazon Web Services Identity and Access Management (Amazon Web Services IAM) policy attached to a resource. These resource shares are visible only to the Amazon Web Services account that created it. They cannot be modified in RAM.

PROMOTING_TO_STANDARD - The resource share is in the process of being promoted. For more information, see PromoteResourceShareCreatedFromPolicy.

STANDARD - Indicates that the resource share was created in RAM using the console or APIs. These resource shares are visible to all principals. They can be modified in RAM.

tags :: Maybe [Tag]

The tags for the resource share.

newResourceShare :: ResourceShare

Create a value of ResourceShare with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ResourceShare', resourceShare_creationTime - The time when the resource share was created.

$sel:status:ResourceShare', resourceShare_status - The status of the resource share.

$sel:resourceShareArn:ResourceShare', resourceShare_resourceShareArn - The Amazon Resource Name (ARN) of the resource share.

$sel:owningAccountId:ResourceShare', resourceShare_owningAccountId - The ID of the Amazon Web Services account that owns the resource share.

$sel:lastUpdatedTime:ResourceShare', resourceShare_lastUpdatedTime - The time when the resource share was last updated.

$sel:allowExternalPrincipals:ResourceShare', resourceShare_allowExternalPrincipals - Indicates whether principals outside your organization in Organizations can be associated with a resource share.

$sel:name:ResourceShare', resourceShare_name - The name of the resource share.

$sel:statusMessage:ResourceShare', resourceShare_statusMessage - A message about the status of the resource share.

$sel:featureSet:ResourceShare', resourceShare_featureSet - Indicates how the resource share was created. Possible values include:

CREATED_FROM_POLICY - Indicates that the resource share was created from an Amazon Web Services Identity and Access Management (Amazon Web Services IAM) policy attached to a resource. These resource shares are visible only to the Amazon Web Services account that created it. They cannot be modified in RAM.

PROMOTING_TO_STANDARD - The resource share is in the process of being promoted. For more information, see PromoteResourceShareCreatedFromPolicy.

STANDARD - Indicates that the resource share was created in RAM using the console or APIs. These resource shares are visible to all principals. They can be modified in RAM.

$sel:tags:ResourceShare', resourceShare_tags - The tags for the resource share.

resourceShare_creationTime :: Lens' ResourceShare (Maybe UTCTime)

The time when the resource share was created.

resourceShare_status :: Lens' ResourceShare (Maybe ResourceShareStatus)

The status of the resource share.

resourceShare_resourceShareArn :: Lens' ResourceShare (Maybe Text)

The Amazon Resource Name (ARN) of the resource share.

resourceShare_owningAccountId :: Lens' ResourceShare (Maybe Text)

The ID of the Amazon Web Services account that owns the resource share.

resourceShare_lastUpdatedTime :: Lens' ResourceShare (Maybe UTCTime)

The time when the resource share was last updated.

resourceShare_allowExternalPrincipals :: Lens' ResourceShare (Maybe Bool)

Indicates whether principals outside your organization in Organizations can be associated with a resource share.

resourceShare_name :: Lens' ResourceShare (Maybe Text)

The name of the resource share.

resourceShare_statusMessage :: Lens' ResourceShare (Maybe Text)

A message about the status of the resource share.

resourceShare_featureSet :: Lens' ResourceShare (Maybe ResourceShareFeatureSet)

Indicates how the resource share was created. Possible values include:

CREATED_FROM_POLICY - Indicates that the resource share was created from an Amazon Web Services Identity and Access Management (Amazon Web Services IAM) policy attached to a resource. These resource shares are visible only to the Amazon Web Services account that created it. They cannot be modified in RAM.

PROMOTING_TO_STANDARD - The resource share is in the process of being promoted. For more information, see PromoteResourceShareCreatedFromPolicy.

STANDARD - Indicates that the resource share was created in RAM using the console or APIs. These resource shares are visible to all principals. They can be modified in RAM.

resourceShare_tags :: Lens' ResourceShare (Maybe [Tag])

The tags for the resource share.

ResourceShareAssociation

data ResourceShareAssociation

Describes an association with a resource share.

See: newResourceShareAssociation smart constructor.

Constructors

ResourceShareAssociation'

Fields

creationTime :: Maybe POSIX

The time when the association was created.

status :: Maybe ResourceShareAssociationStatus

The status of the association.

resourceShareArn :: Maybe Text

The Amazon Resource Name (ARN) of the resource share.

lastUpdatedTime :: Maybe POSIX

The time when the association was last updated.

external :: Maybe Bool

Indicates whether the principal belongs to the same organization in Organizations as the Amazon Web Services account that owns the resource share.

resourceShareName :: Maybe Text

The name of the resource share.

associatedEntity :: Maybe Text

The associated entity. For resource associations, this is the Amazon Resource Name (ARN) of the resource. For principal associations, this is one of the following:

An Amazon Web Services account ID

An ARN of an organization in Organizations

An ARN of an organizational unit (OU) in Organizations

An ARN of an IAM role

An ARN of an IAM user

associationType :: Maybe ResourceShareAssociationType

The association type.

statusMessage :: Maybe Text

A message about the status of the association.

newResourceShareAssociation :: ResourceShareAssociation

Create a value of ResourceShareAssociation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ResourceShareAssociation', resourceShareAssociation_creationTime - The time when the association was created.

$sel:status:ResourceShareAssociation', resourceShareAssociation_status - The status of the association.

$sel:resourceShareArn:ResourceShareAssociation', resourceShareAssociation_resourceShareArn - The Amazon Resource Name (ARN) of the resource share.

$sel:lastUpdatedTime:ResourceShareAssociation', resourceShareAssociation_lastUpdatedTime - The time when the association was last updated.

$sel:external:ResourceShareAssociation', resourceShareAssociation_external - Indicates whether the principal belongs to the same organization in Organizations as the Amazon Web Services account that owns the resource share.

$sel:resourceShareName:ResourceShareAssociation', resourceShareAssociation_resourceShareName - The name of the resource share.

$sel:associatedEntity:ResourceShareAssociation', resourceShareAssociation_associatedEntity - The associated entity. For resource associations, this is the Amazon Resource Name (ARN) of the resource. For principal associations, this is one of the following:

An Amazon Web Services account ID

An ARN of an organization in Organizations

An ARN of an organizational unit (OU) in Organizations

An ARN of an IAM role

An ARN of an IAM user

$sel:associationType:ResourceShareAssociation', resourceShareAssociation_associationType - The association type.

$sel:statusMessage:ResourceShareAssociation', resourceShareAssociation_statusMessage - A message about the status of the association.

resourceShareAssociation_creationTime :: Lens' ResourceShareAssociation (Maybe UTCTime)

The time when the association was created.

resourceShareAssociation_status :: Lens' ResourceShareAssociation (Maybe ResourceShareAssociationStatus)

The status of the association.

resourceShareAssociation_resourceShareArn :: Lens' ResourceShareAssociation (Maybe Text)

The Amazon Resource Name (ARN) of the resource share.

resourceShareAssociation_lastUpdatedTime :: Lens' ResourceShareAssociation (Maybe UTCTime)

The time when the association was last updated.

resourceShareAssociation_external :: Lens' ResourceShareAssociation (Maybe Bool)

Indicates whether the principal belongs to the same organization in Organizations as the Amazon Web Services account that owns the resource share.

resourceShareAssociation_resourceShareName :: Lens' ResourceShareAssociation (Maybe Text)

The name of the resource share.

resourceShareAssociation_associatedEntity :: Lens' ResourceShareAssociation (Maybe Text)

The associated entity. For resource associations, this is the Amazon Resource Name (ARN) of the resource. For principal associations, this is one of the following:

An Amazon Web Services account ID

An ARN of an organization in Organizations

An ARN of an organizational unit (OU) in Organizations

An ARN of an IAM role

An ARN of an IAM user

resourceShareAssociation_associationType :: Lens' ResourceShareAssociation (Maybe ResourceShareAssociationType)

The association type.

resourceShareAssociation_statusMessage :: Lens' ResourceShareAssociation (Maybe Text)

A message about the status of the association.

ResourceShareInvitation

data ResourceShareInvitation

Describes an invitation to join a resource share.

See: newResourceShareInvitation smart constructor.

Constructors

ResourceShareInvitation'

Fields

status :: Maybe ResourceShareInvitationStatus

The status of the invitation.

senderAccountId :: Maybe Text

The ID of the Amazon Web Services account that sent the invitation.

resourceShareArn :: Maybe Text

The Amazon Resource Name (ARN) of the resource share.

receiverAccountId :: Maybe Text

The ID of the Amazon Web Services account that received the invitation.

resourceShareAssociations :: Maybe [ResourceShareAssociation]

To view the resources associated with a pending resource share invitation, use ListPendingInvitationResources.

resourceShareName :: Maybe Text

The name of the resource share.

receiverArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM user or IAM role that received the invitation.

invitationTimestamp :: Maybe POSIX

The date and time when the invitation was sent.

resourceShareInvitationArn :: Maybe Text

The Amazon Resource Name (ARN) of the invitation.

newResourceShareInvitation :: ResourceShareInvitation

Create a value of ResourceShareInvitation with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:ResourceShareInvitation', resourceShareInvitation_status - The status of the invitation.

$sel:senderAccountId:ResourceShareInvitation', resourceShareInvitation_senderAccountId - The ID of the Amazon Web Services account that sent the invitation.

$sel:resourceShareArn:ResourceShareInvitation', resourceShareInvitation_resourceShareArn - The Amazon Resource Name (ARN) of the resource share.

$sel:receiverAccountId:ResourceShareInvitation', resourceShareInvitation_receiverAccountId - The ID of the Amazon Web Services account that received the invitation.

$sel:resourceShareAssociations:ResourceShareInvitation', resourceShareInvitation_resourceShareAssociations - To view the resources associated with a pending resource share invitation, use ListPendingInvitationResources.

$sel:resourceShareName:ResourceShareInvitation', resourceShareInvitation_resourceShareName - The name of the resource share.

$sel:receiverArn:ResourceShareInvitation', resourceShareInvitation_receiverArn - The Amazon Resource Name (ARN) of the IAM user or IAM role that received the invitation.

$sel:invitationTimestamp:ResourceShareInvitation', resourceShareInvitation_invitationTimestamp - The date and time when the invitation was sent.

$sel:resourceShareInvitationArn:ResourceShareInvitation', resourceShareInvitation_resourceShareInvitationArn - The Amazon Resource Name (ARN) of the invitation.

resourceShareInvitation_status :: Lens' ResourceShareInvitation (Maybe ResourceShareInvitationStatus)

The status of the invitation.

resourceShareInvitation_senderAccountId :: Lens' ResourceShareInvitation (Maybe Text)

The ID of the Amazon Web Services account that sent the invitation.

resourceShareInvitation_resourceShareArn :: Lens' ResourceShareInvitation (Maybe Text)

The Amazon Resource Name (ARN) of the resource share.

resourceShareInvitation_receiverAccountId :: Lens' ResourceShareInvitation (Maybe Text)

The ID of the Amazon Web Services account that received the invitation.

resourceShareInvitation_resourceShareAssociations :: Lens' ResourceShareInvitation (Maybe [ResourceShareAssociation])

To view the resources associated with a pending resource share invitation, use ListPendingInvitationResources.

resourceShareInvitation_resourceShareName :: Lens' ResourceShareInvitation (Maybe Text)

The name of the resource share.

resourceShareInvitation_receiverArn :: Lens' ResourceShareInvitation (Maybe Text)

The Amazon Resource Name (ARN) of the IAM user or IAM role that received the invitation.

resourceShareInvitation_invitationTimestamp :: Lens' ResourceShareInvitation (Maybe UTCTime)

The date and time when the invitation was sent.

resourceShareInvitation_resourceShareInvitationArn :: Lens' ResourceShareInvitation (Maybe Text)

The Amazon Resource Name (ARN) of the invitation.

ResourceSharePermissionDetail

data ResourceSharePermissionDetail

Information about an RAM permission.

See: newResourceSharePermissionDetail smart constructor.

Constructors

ResourceSharePermissionDetail'

Fields

creationTime :: Maybe POSIX

The date and time when the permission was created.

resourceType :: Maybe Text

The resource type to which the permission applies.

lastUpdatedTime :: Maybe POSIX

The date and time when the permission was last updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the permission.

defaultVersion :: Maybe Bool

Specifies whether the version of the permission is set to the default version for this permission.

name :: Maybe Text

The name of the permission.

version :: Maybe Text

The identifier for the version of the permission.

isResourceTypeDefault :: Maybe Bool

Specifies whether the version of the permission is set to the default version for this resource type.

permission :: Maybe Text

The permission's effect and actions in JSON format. The effect indicates whether the actions are allowed or denied. The actions list the API actions to which the principal is granted or denied access.

newResourceSharePermissionDetail :: ResourceSharePermissionDetail

Create a value of ResourceSharePermissionDetail with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ResourceSharePermissionDetail', resourceSharePermissionDetail_creationTime - The date and time when the permission was created.

$sel:resourceType:ResourceSharePermissionDetail', resourceSharePermissionDetail_resourceType - The resource type to which the permission applies.

$sel:lastUpdatedTime:ResourceSharePermissionDetail', resourceSharePermissionDetail_lastUpdatedTime - The date and time when the permission was last updated.

$sel:arn:ResourceSharePermissionDetail', resourceSharePermissionDetail_arn - The Amazon Resource Name (ARN) of the permission.

$sel:defaultVersion:ResourceSharePermissionDetail', resourceSharePermissionDetail_defaultVersion - Specifies whether the version of the permission is set to the default version for this permission.

$sel:name:ResourceSharePermissionDetail', resourceSharePermissionDetail_name - The name of the permission.

$sel:version:ResourceSharePermissionDetail', resourceSharePermissionDetail_version - The identifier for the version of the permission.

$sel:isResourceTypeDefault:ResourceSharePermissionDetail', resourceSharePermissionDetail_isResourceTypeDefault - Specifies whether the version of the permission is set to the default version for this resource type.

$sel:permission:ResourceSharePermissionDetail', resourceSharePermissionDetail_permission - The permission's effect and actions in JSON format. The effect indicates whether the actions are allowed or denied. The actions list the API actions to which the principal is granted or denied access.

resourceSharePermissionDetail_creationTime :: Lens' ResourceSharePermissionDetail (Maybe UTCTime)

The date and time when the permission was created.

resourceSharePermissionDetail_resourceType :: Lens' ResourceSharePermissionDetail (Maybe Text)

The resource type to which the permission applies.

resourceSharePermissionDetail_lastUpdatedTime :: Lens' ResourceSharePermissionDetail (Maybe UTCTime)

The date and time when the permission was last updated.

resourceSharePermissionDetail_arn :: Lens' ResourceSharePermissionDetail (Maybe Text)

The Amazon Resource Name (ARN) of the permission.

resourceSharePermissionDetail_defaultVersion :: Lens' ResourceSharePermissionDetail (Maybe Bool)

Specifies whether the version of the permission is set to the default version for this permission.

resourceSharePermissionDetail_name :: Lens' ResourceSharePermissionDetail (Maybe Text)

The name of the permission.

resourceSharePermissionDetail_version :: Lens' ResourceSharePermissionDetail (Maybe Text)

The identifier for the version of the permission.

resourceSharePermissionDetail_isResourceTypeDefault :: Lens' ResourceSharePermissionDetail (Maybe Bool)

Specifies whether the version of the permission is set to the default version for this resource type.

resourceSharePermissionDetail_permission :: Lens' ResourceSharePermissionDetail (Maybe Text)

The permission's effect and actions in JSON format. The effect indicates whether the actions are allowed or denied. The actions list the API actions to which the principal is granted or denied access.

ResourceSharePermissionSummary

data ResourceSharePermissionSummary

Information about a permission that is associated with a resource share.

See: newResourceSharePermissionSummary smart constructor.

Constructors

ResourceSharePermissionSummary'

Fields

creationTime :: Maybe POSIX

The date and time when the permission was created.

status :: Maybe Text

The current status of the permission.

resourceType :: Maybe Text

The type of resource to which the permission applies.

lastUpdatedTime :: Maybe POSIX

The date and time when the permission was last updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the permission.

defaultVersion :: Maybe Bool

Specifies whether the version of the permission is set to the default version for this permission.

name :: Maybe Text

The name of the permission.

version :: Maybe Text

The identifier for the version of the permission.

isResourceTypeDefault :: Maybe Bool

Specifies whether the version of the permission is set to the default version for this resource type.

newResourceSharePermissionSummary :: ResourceSharePermissionSummary

Create a value of ResourceSharePermissionSummary with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:ResourceSharePermissionSummary', resourceSharePermissionSummary_creationTime - The date and time when the permission was created.

$sel:status:ResourceSharePermissionSummary', resourceSharePermissionSummary_status - The current status of the permission.

$sel:resourceType:ResourceSharePermissionSummary', resourceSharePermissionSummary_resourceType - The type of resource to which the permission applies.

$sel:lastUpdatedTime:ResourceSharePermissionSummary', resourceSharePermissionSummary_lastUpdatedTime - The date and time when the permission was last updated.

$sel:arn:ResourceSharePermissionSummary', resourceSharePermissionSummary_arn - The Amazon Resource Name (ARN) of the permission.

$sel:defaultVersion:ResourceSharePermissionSummary', resourceSharePermissionSummary_defaultVersion - Specifies whether the version of the permission is set to the default version for this permission.

$sel:name:ResourceSharePermissionSummary', resourceSharePermissionSummary_name - The name of the permission.

$sel:version:ResourceSharePermissionSummary', resourceSharePermissionSummary_version - The identifier for the version of the permission.

$sel:isResourceTypeDefault:ResourceSharePermissionSummary', resourceSharePermissionSummary_isResourceTypeDefault - Specifies whether the version of the permission is set to the default version for this resource type.

resourceSharePermissionSummary_creationTime :: Lens' ResourceSharePermissionSummary (Maybe UTCTime)

The date and time when the permission was created.

resourceSharePermissionSummary_status :: Lens' ResourceSharePermissionSummary (Maybe Text)

The current status of the permission.

resourceSharePermissionSummary_resourceType :: Lens' ResourceSharePermissionSummary (Maybe Text)

The type of resource to which the permission applies.

resourceSharePermissionSummary_lastUpdatedTime :: Lens' ResourceSharePermissionSummary (Maybe UTCTime)

The date and time when the permission was last updated.

resourceSharePermissionSummary_arn :: Lens' ResourceSharePermissionSummary (Maybe Text)

The Amazon Resource Name (ARN) of the permission.

resourceSharePermissionSummary_defaultVersion :: Lens' ResourceSharePermissionSummary (Maybe Bool)

Specifies whether the version of the permission is set to the default version for this permission.

resourceSharePermissionSummary_name :: Lens' ResourceSharePermissionSummary (Maybe Text)

The name of the permission.

resourceSharePermissionSummary_version :: Lens' ResourceSharePermissionSummary (Maybe Text)

The identifier for the version of the permission.

resourceSharePermissionSummary_isResourceTypeDefault :: Lens' ResourceSharePermissionSummary (Maybe Bool)

Specifies whether the version of the permission is set to the default version for this resource type.

ServiceNameAndResourceType

data ServiceNameAndResourceType

Information about the shareable resource types and the Amazon Web Services services to which they belong.

See: newServiceNameAndResourceType smart constructor.

Constructors

ServiceNameAndResourceType'

Fields

resourceType :: Maybe Text

The shareable resource types.

serviceName :: Maybe Text

The name of the Amazon Web Services services to which the resources belong.

newServiceNameAndResourceType :: ServiceNameAndResourceType

Create a value of ServiceNameAndResourceType with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceType:ServiceNameAndResourceType', serviceNameAndResourceType_resourceType - The shareable resource types.

$sel:serviceName:ServiceNameAndResourceType', serviceNameAndResourceType_serviceName - The name of the Amazon Web Services services to which the resources belong.

serviceNameAndResourceType_resourceType :: Lens' ServiceNameAndResourceType (Maybe Text)

The shareable resource types.

serviceNameAndResourceType_serviceName :: Lens' ServiceNameAndResourceType (Maybe Text)

The name of the Amazon Web Services services to which the resources belong.

Tag

data Tag

Information about a tag.

See: newTag smart constructor.

Constructors

Tag'

Fields

value :: Maybe Text

The value of the tag.

key :: Maybe Text

The key of the tag.

newTag :: Tag

Create a value of Tag with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:Tag', tag_value - The value of the tag.

$sel:key:Tag', tag_key - The key of the tag.

tag_value :: Lens' Tag (Maybe Text)

The value of the tag.

tag_key :: Lens' Tag (Maybe Text)

The key of the tag.

TagFilter

data TagFilter

Used to filter information based on tags.

See: newTagFilter smart constructor.

Constructors

TagFilter'

Fields

tagValues :: Maybe [Text]

The tag values.

tagKey :: Maybe Text

The tag key.

newTagFilter :: TagFilter

Create a value of TagFilter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagValues:TagFilter', tagFilter_tagValues - The tag values.

$sel:tagKey:TagFilter', tagFilter_tagKey - The tag key.

tagFilter_tagValues :: Lens' TagFilter (Maybe [Text])

The tag values.

tagFilter_tagKey :: Lens' TagFilter (Maybe Text)

The tag key.