Network.AWS.RAM.Types.Principal

Description

Documentation

data Principal

Describes a principal for use with Resource Access Manager.

See: newPrincipal smart constructor.

Constructors

Principal'

Fields

creationTime :: Maybe POSIX

The time when the principal was associated with the resource share.

resourceShareArn :: Maybe Text

The Amazon Resource Name (ARN) of the resource share.

lastUpdatedTime :: Maybe POSIX

The time when the association was last updated.

external :: Maybe Bool

Indicates whether the principal belongs to the same organization in Organizations as the Amazon Web Services account that owns the resource share.

id :: Maybe Text

The ID of the principal.

newPrincipal :: Principal

Create a value of Principal with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Principal', principal_creationTime - The time when the principal was associated with the resource share.

$sel:resourceShareArn:Principal', principal_resourceShareArn - The Amazon Resource Name (ARN) of the resource share.

$sel:lastUpdatedTime:Principal', principal_lastUpdatedTime - The time when the association was last updated.

$sel:external:Principal', principal_external - Indicates whether the principal belongs to the same organization in Organizations as the Amazon Web Services account that owns the resource share.

$sel:id:Principal', principal_id - The ID of the principal.

principal_creationTime :: Lens' Principal (Maybe UTCTime)

The time when the principal was associated with the resource share.

principal_resourceShareArn :: Lens' Principal (Maybe Text)

The Amazon Resource Name (ARN) of the resource share.

principal_lastUpdatedTime :: Lens' Principal (Maybe UTCTime)

The time when the association was last updated.

principal_external :: Lens' Principal (Maybe Bool)

Indicates whether the principal belongs to the same organization in Organizations as the Amazon Web Services account that owns the resource share.

principal_id :: Lens' Principal (Maybe Text)

The ID of the principal.