Network.AWS.RAM.Types.Resource

Description

Documentation

data Resource

Describes a resource associated with a resource share.

See: newResource smart constructor.

Constructors

Resource'

Fields

creationTime :: Maybe POSIX

The time when the resource was associated with the resource share.

status :: Maybe ResourceStatus

The status of the resource.

resourceShareArn :: Maybe Text

The Amazon Resource Name (ARN) of the resource share.

lastUpdatedTime :: Maybe POSIX

The time when the association was last updated.

arn :: Maybe Text

The Amazon Resource Name (ARN) of the resource.

resourceGroupArn :: Maybe Text

The Amazon Resource Name (ARN) of the resource group. This value is returned only if the resource is a resource group.

statusMessage :: Maybe Text

A message about the status of the resource.

type' :: Maybe Text

The resource type.

newResource :: Resource

Create a value of Resource with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:creationTime:Resource', resource_creationTime - The time when the resource was associated with the resource share.

$sel:status:Resource', resource_status - The status of the resource.

$sel:resourceShareArn:Resource', resource_resourceShareArn - The Amazon Resource Name (ARN) of the resource share.

$sel:lastUpdatedTime:Resource', resource_lastUpdatedTime - The time when the association was last updated.

$sel:arn:Resource', resource_arn - The Amazon Resource Name (ARN) of the resource.

$sel:resourceGroupArn:Resource', resource_resourceGroupArn - The Amazon Resource Name (ARN) of the resource group. This value is returned only if the resource is a resource group.

$sel:statusMessage:Resource', resource_statusMessage - A message about the status of the resource.

$sel:type':Resource', resource_type - The resource type.

resource_creationTime :: Lens' Resource (Maybe UTCTime)

The time when the resource was associated with the resource share.

resource_status :: Lens' Resource (Maybe ResourceStatus)

The status of the resource.

resource_resourceShareArn :: Lens' Resource (Maybe Text)

The Amazon Resource Name (ARN) of the resource share.

resource_lastUpdatedTime :: Lens' Resource (Maybe UTCTime)

The time when the association was last updated.

resource_arn :: Lens' Resource (Maybe Text)

The Amazon Resource Name (ARN) of the resource.

resource_resourceGroupArn :: Lens' Resource (Maybe Text)

The Amazon Resource Name (ARN) of the resource group. This value is returned only if the resource is a resource group.

resource_statusMessage :: Lens' Resource (Maybe Text)

A message about the status of the resource.

resource_type :: Lens' Resource (Maybe Text)

The resource type.