Network.AWS.RDS.AddRoleToDBCluster

Description

Associates an Identity and Access Management (IAM) role from an Amazon Aurora DB cluster. For more information, see Authorizing Amazon Aurora MySQL to Access Other Amazon Web Services Services on Your Behalf in the Amazon Aurora User Guide.

This action only applies to Aurora DB clusters.

Creating a Request

data AddRoleToDBCluster

See: newAddRoleToDBCluster smart constructor.

Constructors

AddRoleToDBCluster'

Fields

featureName :: Maybe Text

The name of the feature for the DB cluster that the IAM role is to be associated with. For the list of supported feature names, see DBEngineVersion.

dbClusterIdentifier :: Text

The name of the DB cluster to associate the IAM role with.

roleArn :: Text

The Amazon Resource Name (ARN) of the IAM role to associate with the Aurora DB cluster, for example, arn:aws:iam::123456789012:role/AuroraAccessRole.

newAddRoleToDBCluster

Arguments

:: Text
$sel:dbClusterIdentifier:AddRoleToDBCluster'

-> Text
$sel:roleArn:AddRoleToDBCluster'

-> AddRoleToDBCluster

Create a value of AddRoleToDBCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:featureName:AddRoleToDBCluster', addRoleToDBCluster_featureName - The name of the feature for the DB cluster that the IAM role is to be associated with. For the list of supported feature names, see DBEngineVersion.

$sel:dbClusterIdentifier:AddRoleToDBCluster', addRoleToDBCluster_dbClusterIdentifier - The name of the DB cluster to associate the IAM role with.

$sel:roleArn:AddRoleToDBCluster', addRoleToDBCluster_roleArn - The Amazon Resource Name (ARN) of the IAM role to associate with the Aurora DB cluster, for example, arn:aws:iam::123456789012:role/AuroraAccessRole.

Request Lenses

addRoleToDBCluster_featureName :: Lens' AddRoleToDBCluster (Maybe Text)

The name of the feature for the DB cluster that the IAM role is to be associated with. For the list of supported feature names, see DBEngineVersion.

addRoleToDBCluster_dbClusterIdentifier :: Lens' AddRoleToDBCluster Text

The name of the DB cluster to associate the IAM role with.

addRoleToDBCluster_roleArn :: Lens' AddRoleToDBCluster Text

The Amazon Resource Name (ARN) of the IAM role to associate with the Aurora DB cluster, for example, arn:aws:iam::123456789012:role/AuroraAccessRole.

Destructuring the Response

data AddRoleToDBClusterResponse

See: newAddRoleToDBClusterResponse smart constructor.

Constructors

AddRoleToDBClusterResponse'

newAddRoleToDBClusterResponse :: AddRoleToDBClusterResponse

Create a value of AddRoleToDBClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.