Network.AWS.RDS.AuthorizeDBSecurityGroupIngress

Description

Enables ingress to a DBSecurityGroup using one of two forms of authorization. First, EC2 or VPC security groups can be added to the DBSecurityGroup if the application using the database is running on EC2 or VPC instances. Second, IP ranges are available if the application accessing your database is running on the Internet. Required parameters for this API are one of CIDR range, EC2SecurityGroupId for VPC, or (EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId for non-VPC).

You can't authorize ingress from an EC2 security group in one Amazon Web Services Region to an Amazon RDS DB instance in another. You can't authorize ingress from a VPC security group in one VPC to an Amazon RDS DB instance in another.

For an overview of CIDR ranges, go to the Wikipedia Tutorial.

Creating a Request

data AuthorizeDBSecurityGroupIngress

See: newAuthorizeDBSecurityGroupIngress smart constructor.

Constructors

AuthorizeDBSecurityGroupIngress'

Fields

cidrip :: Maybe Text

The IP range to authorize.

eC2SecurityGroupOwnerId :: Maybe Text

Amazon Web Services account number of the owner of the EC2 security group specified in the EC2SecurityGroupName parameter. The Amazon Web Services access key ID isn't an acceptable value. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

eC2SecurityGroupId :: Maybe Text

Id of the EC2 security group to authorize. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

eC2SecurityGroupName :: Maybe Text

Name of the EC2 security group to authorize. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

dbSecurityGroupName :: Text

The name of the DB security group to add authorization to.

newAuthorizeDBSecurityGroupIngress

Arguments

:: Text
$sel:dbSecurityGroupName:AuthorizeDBSecurityGroupIngress'

-> AuthorizeDBSecurityGroupIngress

Create a value of AuthorizeDBSecurityGroupIngress with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cidrip:AuthorizeDBSecurityGroupIngress', authorizeDBSecurityGroupIngress_cidrip - The IP range to authorize.

$sel:eC2SecurityGroupOwnerId:AuthorizeDBSecurityGroupIngress', authorizeDBSecurityGroupIngress_eC2SecurityGroupOwnerId - Amazon Web Services account number of the owner of the EC2 security group specified in the EC2SecurityGroupName parameter. The Amazon Web Services access key ID isn't an acceptable value. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

$sel:eC2SecurityGroupId:AuthorizeDBSecurityGroupIngress', authorizeDBSecurityGroupIngress_eC2SecurityGroupId - Id of the EC2 security group to authorize. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

$sel:eC2SecurityGroupName:AuthorizeDBSecurityGroupIngress', authorizeDBSecurityGroupIngress_eC2SecurityGroupName - Name of the EC2 security group to authorize. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

$sel:dbSecurityGroupName:AuthorizeDBSecurityGroupIngress', authorizeDBSecurityGroupIngress_dbSecurityGroupName - The name of the DB security group to add authorization to.

Request Lenses

authorizeDBSecurityGroupIngress_cidrip :: Lens' AuthorizeDBSecurityGroupIngress (Maybe Text)

The IP range to authorize.

authorizeDBSecurityGroupIngress_eC2SecurityGroupOwnerId :: Lens' AuthorizeDBSecurityGroupIngress (Maybe Text)

Amazon Web Services account number of the owner of the EC2 security group specified in the EC2SecurityGroupName parameter. The Amazon Web Services access key ID isn't an acceptable value. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

authorizeDBSecurityGroupIngress_eC2SecurityGroupId :: Lens' AuthorizeDBSecurityGroupIngress (Maybe Text)

Id of the EC2 security group to authorize. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

authorizeDBSecurityGroupIngress_eC2SecurityGroupName :: Lens' AuthorizeDBSecurityGroupIngress (Maybe Text)

Name of the EC2 security group to authorize. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

authorizeDBSecurityGroupIngress_dbSecurityGroupName :: Lens' AuthorizeDBSecurityGroupIngress Text

The name of the DB security group to add authorization to.

Destructuring the Response

data AuthorizeDBSecurityGroupIngressResponse

See: newAuthorizeDBSecurityGroupIngressResponse smart constructor.

Constructors

AuthorizeDBSecurityGroupIngressResponse'

Fields

dbSecurityGroup :: Maybe DBSecurityGroup

httpStatus :: Int

The response's http status code.

newAuthorizeDBSecurityGroupIngressResponse

Arguments

:: Int
$sel:httpStatus:AuthorizeDBSecurityGroupIngressResponse'

-> AuthorizeDBSecurityGroupIngressResponse

Create a value of AuthorizeDBSecurityGroupIngressResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbSecurityGroup:AuthorizeDBSecurityGroupIngressResponse', authorizeDBSecurityGroupIngressResponse_dbSecurityGroup - Undocumented member.

$sel:httpStatus:AuthorizeDBSecurityGroupIngressResponse', authorizeDBSecurityGroupIngressResponse_httpStatus - The response's http status code.

Response Lenses

authorizeDBSecurityGroupIngressResponse_dbSecurityGroup :: Lens' AuthorizeDBSecurityGroupIngressResponse (Maybe DBSecurityGroup)

Undocumented member.

authorizeDBSecurityGroupIngressResponse_httpStatus :: Lens' AuthorizeDBSecurityGroupIngressResponse Int

The response's http status code.