Network.AWS.RDS.BacktrackDBCluster

Description

Backtracks a DB cluster to a specific time, without creating a new DB cluster.

For more information on backtracking, see Backtracking an Aurora DB Cluster in the Amazon Aurora User Guide.

This action only applies to Aurora MySQL DB clusters.

Creating a Request

data BacktrackDBCluster

See: newBacktrackDBCluster smart constructor.

Constructors

BacktrackDBCluster'

Fields

force :: Maybe Bool

A value that indicates whether to force the DB cluster to backtrack when binary logging is enabled. Otherwise, an error occurs when binary logging is enabled.

useEarliestTimeOnPointInTimeUnavailable :: Maybe Bool

A value that indicates whether to backtrack the DB cluster to the earliest possible backtrack time when BacktrackTo is set to a timestamp earlier than the earliest backtrack time. When this parameter is disabled and BacktrackTo is set to a timestamp earlier than the earliest backtrack time, an error occurs.

dbClusterIdentifier :: Text

The DB cluster identifier of the DB cluster to be backtracked. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: my-cluster1

backtrackTo :: ISO8601

The timestamp of the time to backtrack the DB cluster to, specified in ISO 8601 format. For more information about ISO 8601, see the ISO8601 Wikipedia page.

If the specified time isn't a consistent time for the DB cluster, Aurora automatically chooses the nearest possible consistent time for the DB cluster.

Constraints:

Must contain a valid ISO 8601 timestamp.

Can't contain a timestamp set in the future.

Example: 2017-07-08T18:00Z

newBacktrackDBCluster

Arguments

:: Text
$sel:dbClusterIdentifier:BacktrackDBCluster'

-> UTCTime
$sel:backtrackTo:BacktrackDBCluster'

-> BacktrackDBCluster

Create a value of BacktrackDBCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:force:BacktrackDBCluster', backtrackDBCluster_force - A value that indicates whether to force the DB cluster to backtrack when binary logging is enabled. Otherwise, an error occurs when binary logging is enabled.

$sel:useEarliestTimeOnPointInTimeUnavailable:BacktrackDBCluster', backtrackDBCluster_useEarliestTimeOnPointInTimeUnavailable - A value that indicates whether to backtrack the DB cluster to the earliest possible backtrack time when BacktrackTo is set to a timestamp earlier than the earliest backtrack time. When this parameter is disabled and BacktrackTo is set to a timestamp earlier than the earliest backtrack time, an error occurs.

$sel:dbClusterIdentifier:BacktrackDBCluster', backtrackDBCluster_dbClusterIdentifier - The DB cluster identifier of the DB cluster to be backtracked. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: my-cluster1

$sel:backtrackTo:BacktrackDBCluster', backtrackDBCluster_backtrackTo - The timestamp of the time to backtrack the DB cluster to, specified in ISO 8601 format. For more information about ISO 8601, see the ISO8601 Wikipedia page.

If the specified time isn't a consistent time for the DB cluster, Aurora automatically chooses the nearest possible consistent time for the DB cluster.

Constraints:

Must contain a valid ISO 8601 timestamp.

Can't contain a timestamp set in the future.

Example: 2017-07-08T18:00Z

Request Lenses

backtrackDBCluster_force :: Lens' BacktrackDBCluster (Maybe Bool)

A value that indicates whether to force the DB cluster to backtrack when binary logging is enabled. Otherwise, an error occurs when binary logging is enabled.

backtrackDBCluster_useEarliestTimeOnPointInTimeUnavailable :: Lens' BacktrackDBCluster (Maybe Bool)

A value that indicates whether to backtrack the DB cluster to the earliest possible backtrack time when BacktrackTo is set to a timestamp earlier than the earliest backtrack time. When this parameter is disabled and BacktrackTo is set to a timestamp earlier than the earliest backtrack time, an error occurs.

backtrackDBCluster_dbClusterIdentifier :: Lens' BacktrackDBCluster Text

The DB cluster identifier of the DB cluster to be backtracked. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: my-cluster1

backtrackDBCluster_backtrackTo :: Lens' BacktrackDBCluster UTCTime

The timestamp of the time to backtrack the DB cluster to, specified in ISO 8601 format. For more information about ISO 8601, see the ISO8601 Wikipedia page.

If the specified time isn't a consistent time for the DB cluster, Aurora automatically chooses the nearest possible consistent time for the DB cluster.

Constraints:

Must contain a valid ISO 8601 timestamp.

Can't contain a timestamp set in the future.

Example: 2017-07-08T18:00Z

Destructuring the Response

data DBClusterBacktrack

This data type is used as a response element in the DescribeDBClusterBacktracks action.

See: newDBClusterBacktrack smart constructor.

Constructors

DBClusterBacktrack'

Fields

status :: Maybe Text

The status of the backtrack. This property returns one of the following values:

applying - The backtrack is currently being applied to or rolled back from the DB cluster.

completed - The backtrack has successfully been applied to or rolled back from the DB cluster.

failed - An error occurred while the backtrack was applied to or rolled back from the DB cluster.

pending - The backtrack is currently pending application to or rollback from the DB cluster.

backtrackTo :: Maybe ISO8601

The timestamp of the time to which the DB cluster was backtracked.

dbClusterIdentifier :: Maybe Text

Contains a user-supplied DB cluster identifier. This identifier is the unique key that identifies a DB cluster.

backtrackIdentifier :: Maybe Text

Contains the backtrack identifier.

backtrackRequestCreationTime :: Maybe ISO8601

The timestamp of the time at which the backtrack was requested.

backtrackedFrom :: Maybe ISO8601

The timestamp of the time from which the DB cluster was backtracked.

newDBClusterBacktrack :: DBClusterBacktrack

Create a value of DBClusterBacktrack with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DBClusterBacktrack', dbClusterBacktrack_status - The status of the backtrack. This property returns one of the following values:

applying - The backtrack is currently being applied to or rolled back from the DB cluster.

completed - The backtrack has successfully been applied to or rolled back from the DB cluster.

failed - An error occurred while the backtrack was applied to or rolled back from the DB cluster.

pending - The backtrack is currently pending application to or rollback from the DB cluster.

$sel:backtrackTo:DBClusterBacktrack', dbClusterBacktrack_backtrackTo - The timestamp of the time to which the DB cluster was backtracked.

$sel:dbClusterIdentifier:DBClusterBacktrack', dbClusterBacktrack_dbClusterIdentifier - Contains a user-supplied DB cluster identifier. This identifier is the unique key that identifies a DB cluster.

$sel:backtrackIdentifier:DBClusterBacktrack', dbClusterBacktrack_backtrackIdentifier - Contains the backtrack identifier.

$sel:backtrackRequestCreationTime:DBClusterBacktrack', dbClusterBacktrack_backtrackRequestCreationTime - The timestamp of the time at which the backtrack was requested.

$sel:backtrackedFrom:DBClusterBacktrack', dbClusterBacktrack_backtrackedFrom - The timestamp of the time from which the DB cluster was backtracked.

Response Lenses

dbClusterBacktrack_status :: Lens' DBClusterBacktrack (Maybe Text)

The status of the backtrack. This property returns one of the following values:

applying - The backtrack is currently being applied to or rolled back from the DB cluster.

completed - The backtrack has successfully been applied to or rolled back from the DB cluster.

failed - An error occurred while the backtrack was applied to or rolled back from the DB cluster.

pending - The backtrack is currently pending application to or rollback from the DB cluster.

dbClusterBacktrack_backtrackTo :: Lens' DBClusterBacktrack (Maybe UTCTime)

The timestamp of the time to which the DB cluster was backtracked.

dbClusterBacktrack_dbClusterIdentifier :: Lens' DBClusterBacktrack (Maybe Text)

Contains a user-supplied DB cluster identifier. This identifier is the unique key that identifies a DB cluster.

dbClusterBacktrack_backtrackIdentifier :: Lens' DBClusterBacktrack (Maybe Text)

Contains the backtrack identifier.

dbClusterBacktrack_backtrackRequestCreationTime :: Lens' DBClusterBacktrack (Maybe UTCTime)

The timestamp of the time at which the backtrack was requested.

dbClusterBacktrack_backtrackedFrom :: Lens' DBClusterBacktrack (Maybe UTCTime)

The timestamp of the time from which the DB cluster was backtracked.