Network.AWS.RDS.CopyOptionGroup

Description

Copies the specified option group.

Creating a Request

data CopyOptionGroup

See: newCopyOptionGroup smart constructor.

Constructors

CopyOptionGroup'

Fields

tags :: Maybe [Tag]

sourceOptionGroupIdentifier :: Text

The identifier for the source option group.

Constraints:

Must specify a valid option group.

targetOptionGroupIdentifier :: Text

The identifier for the copied option group.

Constraints:

Can't be null, empty, or blank

Must contain from 1 to 255 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: my-option-group

targetOptionGroupDescription :: Text

The description for the copied option group.

newCopyOptionGroup

Arguments

:: Text
$sel:sourceOptionGroupIdentifier:CopyOptionGroup'

-> Text
$sel:targetOptionGroupIdentifier:CopyOptionGroup'

-> Text
$sel:targetOptionGroupDescription:CopyOptionGroup'

-> CopyOptionGroup

Create a value of CopyOptionGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CopyOptionGroup', copyOptionGroup_tags - Undocumented member.

$sel:sourceOptionGroupIdentifier:CopyOptionGroup', copyOptionGroup_sourceOptionGroupIdentifier - The identifier for the source option group.

Constraints:

Must specify a valid option group.

$sel:targetOptionGroupIdentifier:CopyOptionGroup', copyOptionGroup_targetOptionGroupIdentifier - The identifier for the copied option group.

Constraints:

Can't be null, empty, or blank

Must contain from 1 to 255 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: my-option-group

$sel:targetOptionGroupDescription:CopyOptionGroup', copyOptionGroup_targetOptionGroupDescription - The description for the copied option group.

Request Lenses

copyOptionGroup_tags :: Lens' CopyOptionGroup (Maybe [Tag])

Undocumented member.

copyOptionGroup_sourceOptionGroupIdentifier :: Lens' CopyOptionGroup Text

The identifier for the source option group.

Constraints:

Must specify a valid option group.

copyOptionGroup_targetOptionGroupIdentifier :: Lens' CopyOptionGroup Text

The identifier for the copied option group.

Constraints:

Can't be null, empty, or blank

Must contain from 1 to 255 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: my-option-group

copyOptionGroup_targetOptionGroupDescription :: Lens' CopyOptionGroup Text

The description for the copied option group.

Destructuring the Response

data CopyOptionGroupResponse

See: newCopyOptionGroupResponse smart constructor.

Constructors

CopyOptionGroupResponse'

Fields

optionGroup :: Maybe OptionGroup

httpStatus :: Int

The response's http status code.

newCopyOptionGroupResponse

Arguments

:: Int
$sel:httpStatus:CopyOptionGroupResponse'

-> CopyOptionGroupResponse

Create a value of CopyOptionGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:optionGroup:CopyOptionGroupResponse', copyOptionGroupResponse_optionGroup - Undocumented member.

$sel:httpStatus:CopyOptionGroupResponse', copyOptionGroupResponse_httpStatus - The response's http status code.

Response Lenses

copyOptionGroupResponse_optionGroup :: Lens' CopyOptionGroupResponse (Maybe OptionGroup)

Undocumented member.

copyOptionGroupResponse_httpStatus :: Lens' CopyOptionGroupResponse Int

The response's http status code.