Network.AWS.RDS.CreateDBCluster

Description

Creates a new Amazon Aurora DB cluster.

You can use the ReplicationSourceIdentifier parameter to create the DB cluster as a read replica of another DB cluster or Amazon RDS MySQL or PostgreSQL DB instance. For cross-region replication where the DB cluster identified by ReplicationSourceIdentifier is encrypted, you must also specify the PreSignedUrl parameter.

For more information on Amazon Aurora, see What Is Amazon Aurora? in the Amazon Aurora User Guide.

This action only applies to Aurora DB clusters.

Creating a Request

data CreateDBCluster

See: newCreateDBCluster smart constructor.

Constructors

CreateDBCluster'

Fields

deletionProtection :: Maybe Bool

A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

preferredBackupWindow :: Maybe Text

The daily time range during which automated backups are created if automated backups are enabled using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region. To view the time blocks available, see Backup window in the Amazon Aurora User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

backupRetentionPeriod :: Maybe Int

The number of days for which automated backups are retained.

Default: 1

Constraints:

Must be a value from 1 to 35

storageEncrypted :: Maybe Bool

A value that indicates whether the DB cluster is encrypted.

availabilityZones :: Maybe [Text]

A list of Availability Zones (AZs) where instances in the DB cluster can be created. For information on Amazon Web Services Regions and Availability Zones, see Choosing the Regions and Availability Zones in the Amazon Aurora User Guide.

destinationRegion :: Maybe Text

Pseudo-parameter used when populating the PreSignedUrl of a cross-region CreateDBCluster request. To replicate from region SRC to region DST, send a request to region DST. In that request, pass a PreSignedUrl for region SRC with DestinationRegion set to region DST.

enableIAMDatabaseAuthentication :: Maybe Bool

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information, see IAM Database Authentication in the Amazon Aurora User Guide.

enableCloudwatchLogsExports :: Maybe [Text]

The list of log types that need to be enabled for exporting to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon Aurora User Guide.

Aurora MySQL

Possible values are audit, error, general, and slowquery.

Aurora PostgreSQL

Possible value is postgresql.

optionGroupName :: Maybe Text

A value that indicates that the DB cluster should be associated with the specified option group.

Permanent options can't be removed from an option group. The option group can't be removed from a DB cluster once it is associated with a DB cluster.

domain :: Maybe Text

The Active Directory directory ID to create the DB cluster in.

For Amazon Aurora DB clusters, Amazon RDS can use Kerberos Authentication to authenticate users that connect to the DB cluster. For more information, see Kerberos Authentication in the Amazon Aurora User Guide.

scalingConfiguration :: Maybe ScalingConfiguration

For DB clusters in serverless DB engine mode, the scaling properties of the DB cluster.

dbSubnetGroupName :: Maybe Text

A DB subnet group to associate with this DB cluster.

Constraints: Must match the name of an existing DBSubnetGroup. Must not be default.

Example: mySubnetgroup

engineMode :: Maybe Text

The DB engine mode of the DB cluster, either provisioned, serverless, parallelquery, global, or multimaster.

The parallelquery engine mode isn't required for Aurora MySQL version 1.23 and higher 1.x versions, and version 2.09 and higher 2.x versions.

The global engine mode isn't required for Aurora MySQL version 1.22 and higher 1.x versions, and global engine mode isn't required for any 2.x versions.

The multimaster engine mode only applies for DB clusters created with Aurora MySQL version 5.6.10a.

For Aurora PostgreSQL, the global engine mode isn't required, and both the parallelquery and the multimaster engine modes currently aren't supported.

Limitations and requirements apply to some DB engine modes. For more information, see the following sections in the /Amazon Aurora User Guide/:

Limitations of Aurora Serverless

Limitations of Parallel Query

Limitations of Aurora Global Databases

Limitations of Multi-Master Clusters

masterUserPassword :: Maybe Text

The password for the master database user. This password can contain any printable ASCII character except "/", """, or "@".

Constraints: Must contain from 8 to 41 characters.

masterUsername :: Maybe Text

The name of the master user for the DB cluster.

Constraints:

Must be 1 to 16 letters or numbers.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

enableHttpEndpoint :: Maybe Bool

A value that indicates whether to enable the HTTP endpoint for an Aurora Serverless DB cluster. By default, the HTTP endpoint is disabled.

When enabled, the HTTP endpoint provides a connectionless web service API for running SQL queries on the Aurora Serverless DB cluster. You can also query your database from inside the RDS console with the query editor.

For more information, see Using the Data API for Aurora Serverless in the Amazon Aurora User Guide.

vpcSecurityGroupIds :: Maybe [Text]

A list of EC2 VPC security groups to associate with this DB cluster.

kmsKeyId :: Maybe Text

The Amazon Web Services KMS key identifier for an encrypted DB cluster.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK). To use a CMK in a different Amazon Web Services account, specify the key ARN or alias ARN.

When a CMK isn't specified in KmsKeyId:

If ReplicationSourceIdentifier identifies an encrypted source, then Amazon RDS will use the CMK used to encrypt the source. Otherwise, Amazon RDS will use your default CMK.

If the StorageEncrypted parameter is enabled and ReplicationSourceIdentifier isn't specified, then Amazon RDS will use your default CMK.

There is a default CMK for your Amazon Web Services account. Your Amazon Web Services account has a different default CMK for each Amazon Web Services Region.

If you create a read replica of an encrypted DB cluster in another Amazon Web Services Region, you must set KmsKeyId to a Amazon Web Services KMS key identifier that is valid in the destination Amazon Web Services Region. This CMK is used to encrypt the read replica in that Amazon Web Services Region.

engineVersion :: Maybe Text

The version number of the database engine to use.

To list all of the available engine versions for aurora (for MySQL 5.6-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-mysql (for MySQL 5.7-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora-mysql --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-postgresql, use the following command:

aws rds describe-db-engine-versions --engine aurora-postgresql --query "DBEngineVersions[].EngineVersion"

Aurora MySQL

Example: 5.6.10a, 5.6.mysql_aurora.1.19.2, 5.7.12, 5.7.mysql_aurora.2.04.5

Aurora PostgreSQL

Example: 9.6.3, 10.7

enableGlobalWriteForwarding :: Maybe Bool

A value that indicates whether to enable this DB cluster to forward write operations to the primary cluster of an Aurora global database (GlobalCluster). By default, write operations are not allowed on Aurora DB clusters that are secondary clusters in an Aurora global database.

You can set this value only on Aurora DB clusters that are members of an Aurora global database. With this parameter enabled, a secondary cluster can forward writes to the current primary cluster and the resulting changes are replicated back to this cluster. For the primary DB cluster of an Aurora global database, this value is used immediately if the primary is demoted by the FailoverGlobalCluster API operation, but it does nothing until then.

characterSetName :: Maybe Text

A value that indicates that the DB cluster should be associated with the specified CharacterSet.

preferredMaintenanceWindow :: Maybe Text

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred DB Cluster Maintenance Window in the Amazon Aurora User Guide.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

tags :: Maybe [Tag]

Tags to assign to the DB cluster.

port :: Maybe Int

The port number on which the instances in the DB cluster accept connections.

Default: 3306 if engine is set as aurora or 5432 if set to aurora-postgresql.

domainIAMRoleName :: Maybe Text

Specify the name of the IAM role to be used when making API calls to the Directory Service.

preSignedUrl :: Maybe Text

A URL that contains a Signature Version 4 signed request for the CreateDBCluster action to be called in the source Amazon Web Services Region where the DB cluster is replicated from. You only need to specify PreSignedUrl when you are performing cross-region replication from an encrypted DB cluster.

The pre-signed URL must be a valid request for the CreateDBCluster API action that can be executed in the source Amazon Web Services Region that contains the encrypted DB cluster to be copied.

The pre-signed URL request must contain the following parameter values:

KmsKeyId - The Amazon Web Services KMS key identifier for the key to use to encrypt the copy of the DB cluster in the destination Amazon Web Services Region. This should refer to the same Amazon Web Services KMS CMK for both the CreateDBCluster action that is called in the destination Amazon Web Services Region, and the action contained in the pre-signed URL.

DestinationRegion - The name of the Amazon Web Services Region that Aurora read replica will be created in.

ReplicationSourceIdentifier - The DB cluster identifier for the encrypted DB cluster to be copied. This identifier must be in the Amazon Resource Name (ARN) format for the source Amazon Web Services Region. For example, if you are copying an encrypted DB cluster from the us-west-2 Amazon Web Services Region, then your ReplicationSourceIdentifier would look like Example: arn:aws:rds:us-west-2:123456789012:cluster:aurora-cluster1.

To learn how to generate a Signature Version 4 signed request, see Authenticating Requests: Using Query Parameters (Amazon Web Services Signature Version 4) and Signature Version 4 Signing Process.

If you are using an Amazon Web Services SDK tool or the CLI, you can specify SourceRegion (or --source-region for the CLI) instead of specifying PreSignedUrl manually. Specifying SourceRegion autogenerates a pre-signed URL that is a valid request for the operation that can be executed in the source Amazon Web Services Region.

copyTagsToSnapshot :: Maybe Bool

A value that indicates whether to copy all tags from the DB cluster to snapshots of the DB cluster. The default is not to copy them.

backtrackWindow :: Maybe Integer

The target backtrack window, in seconds. To disable backtracking, set this value to 0.

Currently, Backtrack is only supported for Aurora MySQL DB clusters.

Default: 0

Constraints:

If specified, this value must be set to a number from 0 to 259,200 (72 hours).

dbClusterParameterGroupName :: Maybe Text

The name of the DB cluster parameter group to associate with this DB cluster. If you do not specify a value, then the default DB cluster parameter group for the specified DB engine and version is used.

Constraints:

If supplied, must match the name of an existing DB cluster parameter group.

globalClusterIdentifier :: Maybe Text

The global cluster ID of an Aurora cluster that becomes the primary cluster in the new global database cluster.

replicationSourceIdentifier :: Maybe Text

The Amazon Resource Name (ARN) of the source DB instance or DB cluster if this DB cluster is created as a read replica.

databaseName :: Maybe Text

The name for your database of up to 64 alphanumeric characters. If you do not provide a name, Amazon RDS doesn't create a database in the DB cluster you are creating.

dbClusterIdentifier :: Text

The DB cluster identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: my-cluster1

engine :: Text

The name of the database engine to be used for this DB cluster.

Valid Values: aurora (for MySQL 5.6-compatible Aurora), aurora-mysql (for MySQL 5.7-compatible Aurora), and aurora-postgresql

newCreateDBCluster

Arguments

:: Text
$sel:dbClusterIdentifier:CreateDBCluster'

-> Text
$sel:engine:CreateDBCluster'

-> CreateDBCluster

Create a value of CreateDBCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deletionProtection:CreateDBCluster', createDBCluster_deletionProtection - A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

$sel:preferredBackupWindow:CreateDBCluster', createDBCluster_preferredBackupWindow - The daily time range during which automated backups are created if automated backups are enabled using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region. To view the time blocks available, see Backup window in the Amazon Aurora User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

$sel:backupRetentionPeriod:CreateDBCluster', createDBCluster_backupRetentionPeriod - The number of days for which automated backups are retained.

Default: 1

Constraints:

Must be a value from 1 to 35

$sel:storageEncrypted:CreateDBCluster', createDBCluster_storageEncrypted - A value that indicates whether the DB cluster is encrypted.

$sel:availabilityZones:CreateDBCluster', createDBCluster_availabilityZones - A list of Availability Zones (AZs) where instances in the DB cluster can be created. For information on Amazon Web Services Regions and Availability Zones, see Choosing the Regions and Availability Zones in the Amazon Aurora User Guide.

$sel:destinationRegion:CreateDBCluster', createDBCluster_destinationRegion - Pseudo-parameter used when populating the PreSignedUrl of a cross-region CreateDBCluster request. To replicate from region SRC to region DST, send a request to region DST. In that request, pass a PreSignedUrl for region SRC with DestinationRegion set to region DST.

$sel:enableIAMDatabaseAuthentication:CreateDBCluster', createDBCluster_enableIAMDatabaseAuthentication - A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information, see IAM Database Authentication in the Amazon Aurora User Guide.

$sel:enableCloudwatchLogsExports:CreateDBCluster', createDBCluster_enableCloudwatchLogsExports - The list of log types that need to be enabled for exporting to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon Aurora User Guide.

Aurora MySQL

Possible values are audit, error, general, and slowquery.

Aurora PostgreSQL

Possible value is postgresql.

$sel:optionGroupName:CreateDBCluster', createDBCluster_optionGroupName - A value that indicates that the DB cluster should be associated with the specified option group.

Permanent options can't be removed from an option group. The option group can't be removed from a DB cluster once it is associated with a DB cluster.

$sel:domain:CreateDBCluster', createDBCluster_domain - The Active Directory directory ID to create the DB cluster in.

For Amazon Aurora DB clusters, Amazon RDS can use Kerberos Authentication to authenticate users that connect to the DB cluster. For more information, see Kerberos Authentication in the Amazon Aurora User Guide.

$sel:scalingConfiguration:CreateDBCluster', createDBCluster_scalingConfiguration - For DB clusters in serverless DB engine mode, the scaling properties of the DB cluster.

$sel:dbSubnetGroupName:CreateDBCluster', createDBCluster_dbSubnetGroupName - A DB subnet group to associate with this DB cluster.

Constraints: Must match the name of an existing DBSubnetGroup. Must not be default.

Example: mySubnetgroup

$sel:engineMode:CreateDBCluster', createDBCluster_engineMode - The DB engine mode of the DB cluster, either provisioned, serverless, parallelquery, global, or multimaster.

The parallelquery engine mode isn't required for Aurora MySQL version 1.23 and higher 1.x versions, and version 2.09 and higher 2.x versions.

The global engine mode isn't required for Aurora MySQL version 1.22 and higher 1.x versions, and global engine mode isn't required for any 2.x versions.

The multimaster engine mode only applies for DB clusters created with Aurora MySQL version 5.6.10a.

For Aurora PostgreSQL, the global engine mode isn't required, and both the parallelquery and the multimaster engine modes currently aren't supported.

Limitations and requirements apply to some DB engine modes. For more information, see the following sections in the /Amazon Aurora User Guide/:

Limitations of Aurora Serverless

Limitations of Parallel Query

Limitations of Aurora Global Databases

Limitations of Multi-Master Clusters

$sel:masterUserPassword:CreateDBCluster', createDBCluster_masterUserPassword - The password for the master database user. This password can contain any printable ASCII character except "/", """, or "@".

Constraints: Must contain from 8 to 41 characters.

$sel:masterUsername:CreateDBCluster', createDBCluster_masterUsername - The name of the master user for the DB cluster.

Constraints:

Must be 1 to 16 letters or numbers.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

$sel:enableHttpEndpoint:CreateDBCluster', createDBCluster_enableHttpEndpoint - A value that indicates whether to enable the HTTP endpoint for an Aurora Serverless DB cluster. By default, the HTTP endpoint is disabled.

When enabled, the HTTP endpoint provides a connectionless web service API for running SQL queries on the Aurora Serverless DB cluster. You can also query your database from inside the RDS console with the query editor.

For more information, see Using the Data API for Aurora Serverless in the Amazon Aurora User Guide.

$sel:vpcSecurityGroupIds:CreateDBCluster', createDBCluster_vpcSecurityGroupIds - A list of EC2 VPC security groups to associate with this DB cluster.

$sel:kmsKeyId:CreateDBCluster', createDBCluster_kmsKeyId - The Amazon Web Services KMS key identifier for an encrypted DB cluster.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK). To use a CMK in a different Amazon Web Services account, specify the key ARN or alias ARN.

When a CMK isn't specified in KmsKeyId:

If ReplicationSourceIdentifier identifies an encrypted source, then Amazon RDS will use the CMK used to encrypt the source. Otherwise, Amazon RDS will use your default CMK.

If the StorageEncrypted parameter is enabled and ReplicationSourceIdentifier isn't specified, then Amazon RDS will use your default CMK.

There is a default CMK for your Amazon Web Services account. Your Amazon Web Services account has a different default CMK for each Amazon Web Services Region.

If you create a read replica of an encrypted DB cluster in another Amazon Web Services Region, you must set KmsKeyId to a Amazon Web Services KMS key identifier that is valid in the destination Amazon Web Services Region. This CMK is used to encrypt the read replica in that Amazon Web Services Region.

$sel:engineVersion:CreateDBCluster', createDBCluster_engineVersion - The version number of the database engine to use.

To list all of the available engine versions for aurora (for MySQL 5.6-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-mysql (for MySQL 5.7-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora-mysql --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-postgresql, use the following command:

aws rds describe-db-engine-versions --engine aurora-postgresql --query "DBEngineVersions[].EngineVersion"

Aurora MySQL

Example: 5.6.10a, 5.6.mysql_aurora.1.19.2, 5.7.12, 5.7.mysql_aurora.2.04.5

Aurora PostgreSQL

Example: 9.6.3, 10.7

$sel:enableGlobalWriteForwarding:CreateDBCluster', createDBCluster_enableGlobalWriteForwarding - A value that indicates whether to enable this DB cluster to forward write operations to the primary cluster of an Aurora global database (GlobalCluster). By default, write operations are not allowed on Aurora DB clusters that are secondary clusters in an Aurora global database.

You can set this value only on Aurora DB clusters that are members of an Aurora global database. With this parameter enabled, a secondary cluster can forward writes to the current primary cluster and the resulting changes are replicated back to this cluster. For the primary DB cluster of an Aurora global database, this value is used immediately if the primary is demoted by the FailoverGlobalCluster API operation, but it does nothing until then.

$sel:characterSetName:CreateDBCluster', createDBCluster_characterSetName - A value that indicates that the DB cluster should be associated with the specified CharacterSet.

$sel:preferredMaintenanceWindow:CreateDBCluster', createDBCluster_preferredMaintenanceWindow - The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred DB Cluster Maintenance Window in the Amazon Aurora User Guide.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

$sel:tags:CreateDBCluster', createDBCluster_tags - Tags to assign to the DB cluster.

$sel:port:CreateDBCluster', createDBCluster_port - The port number on which the instances in the DB cluster accept connections.

Default: 3306 if engine is set as aurora or 5432 if set to aurora-postgresql.

$sel:domainIAMRoleName:CreateDBCluster', createDBCluster_domainIAMRoleName - Specify the name of the IAM role to be used when making API calls to the Directory Service.

$sel:preSignedUrl:CreateDBCluster', createDBCluster_preSignedUrl - A URL that contains a Signature Version 4 signed request for the CreateDBCluster action to be called in the source Amazon Web Services Region where the DB cluster is replicated from. You only need to specify PreSignedUrl when you are performing cross-region replication from an encrypted DB cluster.

The pre-signed URL must be a valid request for the CreateDBCluster API action that can be executed in the source Amazon Web Services Region that contains the encrypted DB cluster to be copied.

The pre-signed URL request must contain the following parameter values:

KmsKeyId - The Amazon Web Services KMS key identifier for the key to use to encrypt the copy of the DB cluster in the destination Amazon Web Services Region. This should refer to the same Amazon Web Services KMS CMK for both the CreateDBCluster action that is called in the destination Amazon Web Services Region, and the action contained in the pre-signed URL.

DestinationRegion - The name of the Amazon Web Services Region that Aurora read replica will be created in.

ReplicationSourceIdentifier - The DB cluster identifier for the encrypted DB cluster to be copied. This identifier must be in the Amazon Resource Name (ARN) format for the source Amazon Web Services Region. For example, if you are copying an encrypted DB cluster from the us-west-2 Amazon Web Services Region, then your ReplicationSourceIdentifier would look like Example: arn:aws:rds:us-west-2:123456789012:cluster:aurora-cluster1.

To learn how to generate a Signature Version 4 signed request, see Authenticating Requests: Using Query Parameters (Amazon Web Services Signature Version 4) and Signature Version 4 Signing Process.

If you are using an Amazon Web Services SDK tool or the CLI, you can specify SourceRegion (or --source-region for the CLI) instead of specifying PreSignedUrl manually. Specifying SourceRegion autogenerates a pre-signed URL that is a valid request for the operation that can be executed in the source Amazon Web Services Region.

$sel:copyTagsToSnapshot:CreateDBCluster', createDBCluster_copyTagsToSnapshot - A value that indicates whether to copy all tags from the DB cluster to snapshots of the DB cluster. The default is not to copy them.

$sel:backtrackWindow:CreateDBCluster', createDBCluster_backtrackWindow - The target backtrack window, in seconds. To disable backtracking, set this value to 0.

Currently, Backtrack is only supported for Aurora MySQL DB clusters.

Default: 0

Constraints:

If specified, this value must be set to a number from 0 to 259,200 (72 hours).

$sel:dbClusterParameterGroupName:CreateDBCluster', createDBCluster_dbClusterParameterGroupName - The name of the DB cluster parameter group to associate with this DB cluster. If you do not specify a value, then the default DB cluster parameter group for the specified DB engine and version is used.

Constraints:

If supplied, must match the name of an existing DB cluster parameter group.

$sel:globalClusterIdentifier:CreateDBCluster', createDBCluster_globalClusterIdentifier - The global cluster ID of an Aurora cluster that becomes the primary cluster in the new global database cluster.

$sel:replicationSourceIdentifier:CreateDBCluster', createDBCluster_replicationSourceIdentifier - The Amazon Resource Name (ARN) of the source DB instance or DB cluster if this DB cluster is created as a read replica.

$sel:databaseName:CreateDBCluster', createDBCluster_databaseName - The name for your database of up to 64 alphanumeric characters. If you do not provide a name, Amazon RDS doesn't create a database in the DB cluster you are creating.

$sel:dbClusterIdentifier:CreateDBCluster', createDBCluster_dbClusterIdentifier - The DB cluster identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: my-cluster1

$sel:engine:CreateDBCluster', createDBCluster_engine - The name of the database engine to be used for this DB cluster.

Valid Values: aurora (for MySQL 5.6-compatible Aurora), aurora-mysql (for MySQL 5.7-compatible Aurora), and aurora-postgresql

Request Lenses

createDBCluster_deletionProtection :: Lens' CreateDBCluster (Maybe Bool)

A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

createDBCluster_preferredBackupWindow :: Lens' CreateDBCluster (Maybe Text)

The daily time range during which automated backups are created if automated backups are enabled using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region. To view the time blocks available, see Backup window in the Amazon Aurora User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

createDBCluster_backupRetentionPeriod :: Lens' CreateDBCluster (Maybe Int)

The number of days for which automated backups are retained.

Default: 1

Constraints:

Must be a value from 1 to 35

createDBCluster_storageEncrypted :: Lens' CreateDBCluster (Maybe Bool)

A value that indicates whether the DB cluster is encrypted.

createDBCluster_availabilityZones :: Lens' CreateDBCluster (Maybe [Text])

A list of Availability Zones (AZs) where instances in the DB cluster can be created. For information on Amazon Web Services Regions and Availability Zones, see Choosing the Regions and Availability Zones in the Amazon Aurora User Guide.

createDBCluster_destinationRegion :: Lens' CreateDBCluster (Maybe Text)

Pseudo-parameter used when populating the PreSignedUrl of a cross-region CreateDBCluster request. To replicate from region SRC to region DST, send a request to region DST. In that request, pass a PreSignedUrl for region SRC with DestinationRegion set to region DST.

createDBCluster_enableIAMDatabaseAuthentication :: Lens' CreateDBCluster (Maybe Bool)

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information, see IAM Database Authentication in the Amazon Aurora User Guide.

createDBCluster_enableCloudwatchLogsExports :: Lens' CreateDBCluster (Maybe [Text])

The list of log types that need to be enabled for exporting to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon Aurora User Guide.

Aurora MySQL

Possible values are audit, error, general, and slowquery.

Aurora PostgreSQL

Possible value is postgresql.

createDBCluster_optionGroupName :: Lens' CreateDBCluster (Maybe Text)

A value that indicates that the DB cluster should be associated with the specified option group.

Permanent options can't be removed from an option group. The option group can't be removed from a DB cluster once it is associated with a DB cluster.

createDBCluster_domain :: Lens' CreateDBCluster (Maybe Text)

The Active Directory directory ID to create the DB cluster in.

For Amazon Aurora DB clusters, Amazon RDS can use Kerberos Authentication to authenticate users that connect to the DB cluster. For more information, see Kerberos Authentication in the Amazon Aurora User Guide.

createDBCluster_scalingConfiguration :: Lens' CreateDBCluster (Maybe ScalingConfiguration)

For DB clusters in serverless DB engine mode, the scaling properties of the DB cluster.

createDBCluster_dbSubnetGroupName :: Lens' CreateDBCluster (Maybe Text)

A DB subnet group to associate with this DB cluster.

Constraints: Must match the name of an existing DBSubnetGroup. Must not be default.

Example: mySubnetgroup

createDBCluster_engineMode :: Lens' CreateDBCluster (Maybe Text)

The DB engine mode of the DB cluster, either provisioned, serverless, parallelquery, global, or multimaster.

The parallelquery engine mode isn't required for Aurora MySQL version 1.23 and higher 1.x versions, and version 2.09 and higher 2.x versions.

The global engine mode isn't required for Aurora MySQL version 1.22 and higher 1.x versions, and global engine mode isn't required for any 2.x versions.

The multimaster engine mode only applies for DB clusters created with Aurora MySQL version 5.6.10a.

For Aurora PostgreSQL, the global engine mode isn't required, and both the parallelquery and the multimaster engine modes currently aren't supported.

Limitations and requirements apply to some DB engine modes. For more information, see the following sections in the /Amazon Aurora User Guide/:

Limitations of Aurora Serverless

Limitations of Parallel Query

Limitations of Aurora Global Databases

Limitations of Multi-Master Clusters

createDBCluster_masterUserPassword :: Lens' CreateDBCluster (Maybe Text)

The password for the master database user. This password can contain any printable ASCII character except "/", """, or "@".

Constraints: Must contain from 8 to 41 characters.

createDBCluster_masterUsername :: Lens' CreateDBCluster (Maybe Text)

The name of the master user for the DB cluster.

Constraints:

Must be 1 to 16 letters or numbers.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

createDBCluster_enableHttpEndpoint :: Lens' CreateDBCluster (Maybe Bool)

A value that indicates whether to enable the HTTP endpoint for an Aurora Serverless DB cluster. By default, the HTTP endpoint is disabled.

When enabled, the HTTP endpoint provides a connectionless web service API for running SQL queries on the Aurora Serverless DB cluster. You can also query your database from inside the RDS console with the query editor.

For more information, see Using the Data API for Aurora Serverless in the Amazon Aurora User Guide.

createDBCluster_vpcSecurityGroupIds :: Lens' CreateDBCluster (Maybe [Text])

A list of EC2 VPC security groups to associate with this DB cluster.

createDBCluster_kmsKeyId :: Lens' CreateDBCluster (Maybe Text)

The Amazon Web Services KMS key identifier for an encrypted DB cluster.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK). To use a CMK in a different Amazon Web Services account, specify the key ARN or alias ARN.

When a CMK isn't specified in KmsKeyId:

If ReplicationSourceIdentifier identifies an encrypted source, then Amazon RDS will use the CMK used to encrypt the source. Otherwise, Amazon RDS will use your default CMK.

If the StorageEncrypted parameter is enabled and ReplicationSourceIdentifier isn't specified, then Amazon RDS will use your default CMK.

There is a default CMK for your Amazon Web Services account. Your Amazon Web Services account has a different default CMK for each Amazon Web Services Region.

If you create a read replica of an encrypted DB cluster in another Amazon Web Services Region, you must set KmsKeyId to a Amazon Web Services KMS key identifier that is valid in the destination Amazon Web Services Region. This CMK is used to encrypt the read replica in that Amazon Web Services Region.

createDBCluster_engineVersion :: Lens' CreateDBCluster (Maybe Text)

The version number of the database engine to use.

To list all of the available engine versions for aurora (for MySQL 5.6-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-mysql (for MySQL 5.7-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora-mysql --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-postgresql, use the following command:

aws rds describe-db-engine-versions --engine aurora-postgresql --query "DBEngineVersions[].EngineVersion"

Aurora MySQL

Example: 5.6.10a, 5.6.mysql_aurora.1.19.2, 5.7.12, 5.7.mysql_aurora.2.04.5

Aurora PostgreSQL

Example: 9.6.3, 10.7

createDBCluster_enableGlobalWriteForwarding :: Lens' CreateDBCluster (Maybe Bool)

A value that indicates whether to enable this DB cluster to forward write operations to the primary cluster of an Aurora global database (GlobalCluster). By default, write operations are not allowed on Aurora DB clusters that are secondary clusters in an Aurora global database.

You can set this value only on Aurora DB clusters that are members of an Aurora global database. With this parameter enabled, a secondary cluster can forward writes to the current primary cluster and the resulting changes are replicated back to this cluster. For the primary DB cluster of an Aurora global database, this value is used immediately if the primary is demoted by the FailoverGlobalCluster API operation, but it does nothing until then.

createDBCluster_characterSetName :: Lens' CreateDBCluster (Maybe Text)

A value that indicates that the DB cluster should be associated with the specified CharacterSet.

createDBCluster_preferredMaintenanceWindow :: Lens' CreateDBCluster (Maybe Text)

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred DB Cluster Maintenance Window in the Amazon Aurora User Guide.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

createDBCluster_tags :: Lens' CreateDBCluster (Maybe [Tag])

Tags to assign to the DB cluster.

createDBCluster_port :: Lens' CreateDBCluster (Maybe Int)

The port number on which the instances in the DB cluster accept connections.

Default: 3306 if engine is set as aurora or 5432 if set to aurora-postgresql.

createDBCluster_domainIAMRoleName :: Lens' CreateDBCluster (Maybe Text)

Specify the name of the IAM role to be used when making API calls to the Directory Service.

createDBCluster_preSignedUrl :: Lens' CreateDBCluster (Maybe Text)

A URL that contains a Signature Version 4 signed request for the CreateDBCluster action to be called in the source Amazon Web Services Region where the DB cluster is replicated from. You only need to specify PreSignedUrl when you are performing cross-region replication from an encrypted DB cluster.

The pre-signed URL must be a valid request for the CreateDBCluster API action that can be executed in the source Amazon Web Services Region that contains the encrypted DB cluster to be copied.

The pre-signed URL request must contain the following parameter values:

KmsKeyId - The Amazon Web Services KMS key identifier for the key to use to encrypt the copy of the DB cluster in the destination Amazon Web Services Region. This should refer to the same Amazon Web Services KMS CMK for both the CreateDBCluster action that is called in the destination Amazon Web Services Region, and the action contained in the pre-signed URL.

DestinationRegion - The name of the Amazon Web Services Region that Aurora read replica will be created in.

ReplicationSourceIdentifier - The DB cluster identifier for the encrypted DB cluster to be copied. This identifier must be in the Amazon Resource Name (ARN) format for the source Amazon Web Services Region. For example, if you are copying an encrypted DB cluster from the us-west-2 Amazon Web Services Region, then your ReplicationSourceIdentifier would look like Example: arn:aws:rds:us-west-2:123456789012:cluster:aurora-cluster1.

To learn how to generate a Signature Version 4 signed request, see Authenticating Requests: Using Query Parameters (Amazon Web Services Signature Version 4) and Signature Version 4 Signing Process.

If you are using an Amazon Web Services SDK tool or the CLI, you can specify SourceRegion (or --source-region for the CLI) instead of specifying PreSignedUrl manually. Specifying SourceRegion autogenerates a pre-signed URL that is a valid request for the operation that can be executed in the source Amazon Web Services Region.

createDBCluster_copyTagsToSnapshot :: Lens' CreateDBCluster (Maybe Bool)

A value that indicates whether to copy all tags from the DB cluster to snapshots of the DB cluster. The default is not to copy them.

createDBCluster_backtrackWindow :: Lens' CreateDBCluster (Maybe Integer)

The target backtrack window, in seconds. To disable backtracking, set this value to 0.

Currently, Backtrack is only supported for Aurora MySQL DB clusters.

Default: 0

Constraints:

If specified, this value must be set to a number from 0 to 259,200 (72 hours).

createDBCluster_dbClusterParameterGroupName :: Lens' CreateDBCluster (Maybe Text)

The name of the DB cluster parameter group to associate with this DB cluster. If you do not specify a value, then the default DB cluster parameter group for the specified DB engine and version is used.

Constraints:

If supplied, must match the name of an existing DB cluster parameter group.

createDBCluster_globalClusterIdentifier :: Lens' CreateDBCluster (Maybe Text)

The global cluster ID of an Aurora cluster that becomes the primary cluster in the new global database cluster.

createDBCluster_replicationSourceIdentifier :: Lens' CreateDBCluster (Maybe Text)

The Amazon Resource Name (ARN) of the source DB instance or DB cluster if this DB cluster is created as a read replica.

createDBCluster_databaseName :: Lens' CreateDBCluster (Maybe Text)

The name for your database of up to 64 alphanumeric characters. If you do not provide a name, Amazon RDS doesn't create a database in the DB cluster you are creating.

createDBCluster_dbClusterIdentifier :: Lens' CreateDBCluster Text

The DB cluster identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: my-cluster1

createDBCluster_engine :: Lens' CreateDBCluster Text

The name of the database engine to be used for this DB cluster.

Valid Values: aurora (for MySQL 5.6-compatible Aurora), aurora-mysql (for MySQL 5.7-compatible Aurora), and aurora-postgresql

Destructuring the Response

data CreateDBClusterResponse

See: newCreateDBClusterResponse smart constructor.

Constructors

CreateDBClusterResponse'

Fields

dbCluster :: Maybe DBCluster

httpStatus :: Int

The response's http status code.

newCreateDBClusterResponse

Arguments

:: Int
$sel:httpStatus:CreateDBClusterResponse'

-> CreateDBClusterResponse

Create a value of CreateDBClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbCluster:CreateDBClusterResponse', createDBClusterResponse_dbCluster - Undocumented member.

$sel:httpStatus:CreateDBClusterResponse', createDBClusterResponse_httpStatus - The response's http status code.

Response Lenses

createDBClusterResponse_dbCluster :: Lens' CreateDBClusterResponse (Maybe DBCluster)

Undocumented member.

createDBClusterResponse_httpStatus :: Lens' CreateDBClusterResponse Int

The response's http status code.