Network.AWS.RDS.CreateDBInstanceReadReplica

Description

Creates a new DB instance that acts as a read replica for an existing source DB instance. You can create a read replica for a DB instance running MySQL, MariaDB, Oracle, PostgreSQL, or SQL Server. For more information, see Working with Read Replicas in the Amazon RDS User Guide.

Amazon Aurora doesn't support this action. Call the CreateDBInstance action to create a DB instance for an Aurora DB cluster.

All read replica DB instances are created with backups disabled. All other DB instance attributes (including DB security groups and DB parameter groups) are inherited from the source DB instance, except as specified.

Your source DB instance must have backup retention enabled.

Creating a Request

data CreateDBInstanceReadReplica

See: newCreateDBInstanceReadReplica smart constructor.

Constructors

CreateDBInstanceReadReplica'

Fields

deletionProtection :: Maybe Bool

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

enablePerformanceInsights :: Maybe Bool

A value that indicates whether to enable Performance Insights for the read replica.

For more information, see Using Amazon Performance Insights in the Amazon RDS User Guide.

destinationRegion :: Maybe Text

Pseudo-parameter used when populating the PreSignedUrl of a cross-region CreateDBInstanceReadReplica request. To replicate from region SRC to region DST, send a request to region DST. In that request, pass a PreSignedUrl for region SRC with DestinationRegion set to region DST.

maxAllocatedStorage :: Maybe Int

The upper limit in gibibytes (GiB) to which Amazon RDS can automatically scale the storage of the DB instance.

For more information about this setting, including limitations that apply to it, see Managing capacity automatically with Amazon RDS storage autoscaling in the Amazon RDS User Guide.

enableIAMDatabaseAuthentication :: Maybe Bool

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information about IAM database authentication, see IAM Database Authentication for MySQL and PostgreSQL in the Amazon RDS User Guide.

useDefaultProcessorFeatures :: Maybe Bool

A value that indicates whether the DB instance class of the DB instance uses its default processor features.

enableCloudwatchLogsExports :: Maybe [Text]

The list of logs that the new DB instance is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon RDS User Guide.

storageType :: Maybe Text

Specifies the storage type to be associated with the read replica.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified, otherwise gp2

optionGroupName :: Maybe Text

The option group the DB instance is associated with. If omitted, the option group associated with the source instance is used.

For SQL Server, you must use the option group associated with the source instance.

monitoringInterval :: Maybe Int

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the read replica. To disable collecting Enhanced Monitoring metrics, specify 0. The default is 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

domain :: Maybe Text

The Active Directory directory ID to create the DB instance in. Currently, only MySQL, Microsoft SQL Server, Oracle, and PostgreSQL DB instances can be created in an Active Directory Domain.

For more information, see Kerberos Authentication in the Amazon RDS User Guide.

monitoringRoleArn :: Maybe Text

The ARN for the IAM role that permits RDS to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess. For information on creating a monitoring role, go to To create an IAM role for Amazon RDS Enhanced Monitoring in the Amazon RDS User Guide.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

dbSubnetGroupName :: Maybe Text

Specifies a DB subnet group for the DB instance. The new DB instance is created in the VPC associated with the DB subnet group. If no DB subnet group is specified, then the new DB instance isn't created in a VPC.

Constraints:

Can only be specified if the source DB instance identifier specifies a DB instance in another Amazon Web Services Region.

If supplied, must match the name of an existing DBSubnetGroup.

The specified DB subnet group must be in the same Amazon Web Services Region in which the operation is running.

All read replicas in one Amazon Web Services Region that are created from the same source DB instance must either:>

Specify DB subnet groups from the same VPC. All these read replicas are created in the same VPC.

Not specify a DB subnet group. All these read replicas are created outside of any VPC.

Example: mySubnetgroup

publiclyAccessible :: Maybe Bool

A value that indicates whether the DB instance is publicly accessible.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

For more information, see CreateDBInstance.

multiAZ :: Maybe Bool

A value that indicates whether the read replica is in a Multi-AZ deployment.

You can create a read replica as a Multi-AZ DB instance. RDS creates a standby of your replica in another Availability Zone for failover support for the replica. Creating your read replica as a Multi-AZ DB instance is independent of whether the source database is a Multi-AZ DB instance.

performanceInsightsKMSKeyId :: Maybe Text

The Amazon Web Services KMS key identifier for encryption of Performance Insights data.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

If you do not specify a value for PerformanceInsightsKMSKeyId, then Amazon RDS uses your default CMK. There is a default CMK for your Amazon Web Services account. Your Amazon Web Services account has a different default CMK for each Amazon Web Services Region.

vpcSecurityGroupIds :: Maybe [Text]

A list of EC2 VPC security groups to associate with the read replica.

Default: The default EC2 VPC security group for the DB subnet group's VPC.

kmsKeyId :: Maybe Text

The Amazon Web Services KMS key identifier for an encrypted read replica.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS CMK.

If you create an encrypted read replica in the same Amazon Web Services Region as the source DB instance, then do not specify a value for this parameter. A read replica in the same Region is always encrypted with the same Amazon Web Services KMS CMK as the source DB instance.

If you create an encrypted read replica in a different Amazon Web Services Region, then you must specify a Amazon Web Services KMS key identifier for the destination Amazon Web Services Region. Amazon Web Services KMS CMKs are specific to the Amazon Web Services Region that they are created in, and you can't use CMKs from one Amazon Web Services Region in another Amazon Web Services Region.

You can't create an encrypted read replica from an unencrypted DB instance.

dbParameterGroupName :: Maybe Text

The name of the DB parameter group to associate with this DB instance.

If you do not specify a value for DBParameterGroupName, then Amazon RDS uses the DBParameterGroup of source DB instance for a same region read replica, or the default DBParameterGroup for the specified DB engine for a cross region read replica.

Currently, specifying a parameter group for this operation is only supported for Oracle DB instances.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

availabilityZone :: Maybe Text

The Availability Zone (AZ) where the read replica will be created.

Default: A random, system-chosen Availability Zone in the endpoint's Amazon Web Services Region.

Example: us-east-1d

performanceInsightsRetentionPeriod :: Maybe Int

The amount of time, in days, to retain Performance Insights data. Valid values are 7 or 731 (2 years).

processorFeatures :: Maybe [ProcessorFeature]

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

tags :: Maybe [Tag]

dbInstanceClass :: Maybe Text

The compute and memory capacity of the read replica, for example, db.m4.large. Not all DB instance classes are available in all Amazon Web Services Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

Default: Inherits from the source DB instance.

port :: Maybe Int

The port number that the DB instance uses for connections.

Default: Inherits from the source DB instance

Valid Values: 1150-65535

domainIAMRoleName :: Maybe Text

Specify the name of the IAM role to be used when making API calls to the Directory Service.

preSignedUrl :: Maybe Text

The URL that contains a Signature Version 4 signed request for the CreateDBInstanceReadReplica API action in the source Amazon Web Services Region that contains the source DB instance.

You must specify this parameter when you create an encrypted read replica from another Amazon Web Services Region by using the Amazon RDS API. Don't specify PreSignedUrl when you are creating an encrypted read replica in the same Amazon Web Services Region.

The presigned URL must be a valid request for the CreateDBInstanceReadReplica API action that can be executed in the source Amazon Web Services Region that contains the encrypted source DB instance. The presigned URL request must contain the following parameter values:

DestinationRegion - The Amazon Web Services Region that the encrypted read replica is created in. This Amazon Web Services Region is the same one where the CreateDBInstanceReadReplica action is called that contains this presigned URL.

For example, if you create an encrypted DB instance in the us-west-1 Amazon Web Services Region, from a source DB instance in the us-east-2 Amazon Web Services Region, then you call the CreateDBInstanceReadReplica action in the us-east-1 Amazon Web Services Region and provide a presigned URL that contains a call to the CreateDBInstanceReadReplica action in the us-west-2 Amazon Web Services Region. For this example, the DestinationRegion in the presigned URL must be set to the us-east-1 Amazon Web Services Region.

KmsKeyId - The Amazon Web Services KMS key identifier for the key to use to encrypt the read replica in the destination Amazon Web Services Region. This is the same identifier for both the CreateDBInstanceReadReplica action that is called in the destination Amazon Web Services Region, and the action contained in the presigned URL.

SourceDBInstanceIdentifier - The DB instance identifier for the encrypted DB instance to be replicated. This identifier must be in the Amazon Resource Name (ARN) format for the source Amazon Web Services Region. For example, if you are creating an encrypted read replica from a DB instance in the us-west-2 Amazon Web Services Region, then your SourceDBInstanceIdentifier looks like the following example: arn:aws:rds:us-west-2:123456789012:instance:mysql-instance1-20161115.

To learn how to generate a Signature Version 4 signed request, see Authenticating Requests: Using Query Parameters (Amazon Web Services Signature Version 4) and Signature Version 4 Signing Process.

If you are using an Amazon Web Services SDK tool or the CLI, you can specify SourceRegion (or --source-region for the CLI) instead of specifying PreSignedUrl manually. Specifying SourceRegion autogenerates a presigned URL that is a valid request for the operation that can be executed in the source Amazon Web Services Region.

SourceRegion isn't supported for SQL Server, because SQL Server on Amazon RDS doesn't support cross-region read replicas.

replicaMode :: Maybe ReplicaMode

The open mode of the replica database: mounted or read-only.

This parameter is only supported for Oracle DB instances.

Mounted DB replicas are included in Oracle Enterprise Edition. The main use case for mounted replicas is cross-Region disaster recovery. The primary database doesn't use Active Data Guard to transmit information to the mounted replica. Because it doesn't accept user connections, a mounted replica can't serve a read-only workload.

You can create a combination of mounted and read-only DB replicas for the same primary DB instance. For more information, see Working with Oracle Read Replicas for Amazon RDS in the Amazon RDS User Guide.

copyTagsToSnapshot :: Maybe Bool

A value that indicates whether to copy all tags from the read replica to snapshots of the read replica. By default, tags are not copied.

iops :: Maybe Int

The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance.

autoMinorVersionUpgrade :: Maybe Bool

A value that indicates whether minor engine upgrades are applied automatically to the read replica during the maintenance window.

Default: Inherits from the source DB instance

dbInstanceIdentifier :: Text

The DB instance identifier of the read replica. This identifier is the unique key that identifies a DB instance. This parameter is stored as a lowercase string.

sourceDBInstanceIdentifier :: Text

The identifier of the DB instance that will act as the source for the read replica. Each DB instance can have up to five read replicas.

Constraints:

Must be the identifier of an existing MySQL, MariaDB, Oracle, PostgreSQL, or SQL Server DB instance.

Can specify a DB instance that is a MySQL read replica only if the source is running MySQL 5.6 or later.

For the limitations of Oracle read replicas, see Read Replica Limitations with Oracle in the Amazon RDS User Guide.

For the limitations of SQL Server read replicas, see Read Replica Limitations with Microsoft SQL Server in the Amazon RDS User Guide.

Can specify a PostgreSQL DB instance only if the source is running PostgreSQL 9.3.5 or later (9.4.7 and higher for cross-region replication).

The specified DB instance must have automatic backups enabled, that is, its backup retention period must be greater than 0.

If the source DB instance is in the same Amazon Web Services Region as the read replica, specify a valid DB instance identifier.

If the source DB instance is in a different Amazon Web Services Region from the read replica, specify a valid DB instance ARN. For more information, see Constructing an ARN for Amazon RDS in the Amazon RDS User Guide. This doesn't apply to SQL Server, which doesn't support cross-region replicas.

newCreateDBInstanceReadReplica

Arguments

:: Text
$sel:dbInstanceIdentifier:CreateDBInstanceReadReplica'

-> Text
$sel:sourceDBInstanceIdentifier:CreateDBInstanceReadReplica'

-> CreateDBInstanceReadReplica

Create a value of CreateDBInstanceReadReplica with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deletionProtection:CreateDBInstanceReadReplica', createDBInstanceReadReplica_deletionProtection - A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

$sel:enablePerformanceInsights:CreateDBInstanceReadReplica', createDBInstanceReadReplica_enablePerformanceInsights - A value that indicates whether to enable Performance Insights for the read replica.

For more information, see Using Amazon Performance Insights in the Amazon RDS User Guide.

$sel:destinationRegion:CreateDBInstanceReadReplica', createDBInstanceReadReplica_destinationRegion - Pseudo-parameter used when populating the PreSignedUrl of a cross-region CreateDBInstanceReadReplica request. To replicate from region SRC to region DST, send a request to region DST. In that request, pass a PreSignedUrl for region SRC with DestinationRegion set to region DST.

$sel:maxAllocatedStorage:CreateDBInstanceReadReplica', createDBInstanceReadReplica_maxAllocatedStorage - The upper limit in gibibytes (GiB) to which Amazon RDS can automatically scale the storage of the DB instance.

For more information about this setting, including limitations that apply to it, see Managing capacity automatically with Amazon RDS storage autoscaling in the Amazon RDS User Guide.

$sel:enableIAMDatabaseAuthentication:CreateDBInstanceReadReplica', createDBInstanceReadReplica_enableIAMDatabaseAuthentication - A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information about IAM database authentication, see IAM Database Authentication for MySQL and PostgreSQL in the Amazon RDS User Guide.

$sel:useDefaultProcessorFeatures:CreateDBInstanceReadReplica', createDBInstanceReadReplica_useDefaultProcessorFeatures - A value that indicates whether the DB instance class of the DB instance uses its default processor features.

$sel:enableCloudwatchLogsExports:CreateDBInstanceReadReplica', createDBInstanceReadReplica_enableCloudwatchLogsExports - The list of logs that the new DB instance is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon RDS User Guide.

$sel:storageType:CreateDBInstanceReadReplica', createDBInstanceReadReplica_storageType - Specifies the storage type to be associated with the read replica.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified, otherwise gp2

$sel:optionGroupName:CreateDBInstanceReadReplica', createDBInstanceReadReplica_optionGroupName - The option group the DB instance is associated with. If omitted, the option group associated with the source instance is used.

For SQL Server, you must use the option group associated with the source instance.

$sel:monitoringInterval:CreateDBInstanceReadReplica', createDBInstanceReadReplica_monitoringInterval - The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the read replica. To disable collecting Enhanced Monitoring metrics, specify 0. The default is 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

$sel:domain:CreateDBInstanceReadReplica', createDBInstanceReadReplica_domain - The Active Directory directory ID to create the DB instance in. Currently, only MySQL, Microsoft SQL Server, Oracle, and PostgreSQL DB instances can be created in an Active Directory Domain.

For more information, see Kerberos Authentication in the Amazon RDS User Guide.

$sel:monitoringRoleArn:CreateDBInstanceReadReplica', createDBInstanceReadReplica_monitoringRoleArn - The ARN for the IAM role that permits RDS to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess. For information on creating a monitoring role, go to To create an IAM role for Amazon RDS Enhanced Monitoring in the Amazon RDS User Guide.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

$sel:dbSubnetGroupName:CreateDBInstanceReadReplica', createDBInstanceReadReplica_dbSubnetGroupName - Specifies a DB subnet group for the DB instance. The new DB instance is created in the VPC associated with the DB subnet group. If no DB subnet group is specified, then the new DB instance isn't created in a VPC.

Constraints:

Can only be specified if the source DB instance identifier specifies a DB instance in another Amazon Web Services Region.

If supplied, must match the name of an existing DBSubnetGroup.

The specified DB subnet group must be in the same Amazon Web Services Region in which the operation is running.

All read replicas in one Amazon Web Services Region that are created from the same source DB instance must either:>

Specify DB subnet groups from the same VPC. All these read replicas are created in the same VPC.

Not specify a DB subnet group. All these read replicas are created outside of any VPC.

Example: mySubnetgroup

$sel:publiclyAccessible:CreateDBInstanceReadReplica', createDBInstanceReadReplica_publiclyAccessible - A value that indicates whether the DB instance is publicly accessible.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

For more information, see CreateDBInstance.

$sel:multiAZ:CreateDBInstanceReadReplica', createDBInstanceReadReplica_multiAZ - A value that indicates whether the read replica is in a Multi-AZ deployment.

You can create a read replica as a Multi-AZ DB instance. RDS creates a standby of your replica in another Availability Zone for failover support for the replica. Creating your read replica as a Multi-AZ DB instance is independent of whether the source database is a Multi-AZ DB instance.

$sel:performanceInsightsKMSKeyId:CreateDBInstanceReadReplica', createDBInstanceReadReplica_performanceInsightsKMSKeyId - The Amazon Web Services KMS key identifier for encryption of Performance Insights data.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

If you do not specify a value for PerformanceInsightsKMSKeyId, then Amazon RDS uses your default CMK. There is a default CMK for your Amazon Web Services account. Your Amazon Web Services account has a different default CMK for each Amazon Web Services Region.

$sel:vpcSecurityGroupIds:CreateDBInstanceReadReplica', createDBInstanceReadReplica_vpcSecurityGroupIds - A list of EC2 VPC security groups to associate with the read replica.

Default: The default EC2 VPC security group for the DB subnet group's VPC.

$sel:kmsKeyId:CreateDBInstanceReadReplica', createDBInstanceReadReplica_kmsKeyId - The Amazon Web Services KMS key identifier for an encrypted read replica.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS CMK.

If you create an encrypted read replica in the same Amazon Web Services Region as the source DB instance, then do not specify a value for this parameter. A read replica in the same Region is always encrypted with the same Amazon Web Services KMS CMK as the source DB instance.

If you create an encrypted read replica in a different Amazon Web Services Region, then you must specify a Amazon Web Services KMS key identifier for the destination Amazon Web Services Region. Amazon Web Services KMS CMKs are specific to the Amazon Web Services Region that they are created in, and you can't use CMKs from one Amazon Web Services Region in another Amazon Web Services Region.

You can't create an encrypted read replica from an unencrypted DB instance.

$sel:dbParameterGroupName:CreateDBInstanceReadReplica', createDBInstanceReadReplica_dbParameterGroupName - The name of the DB parameter group to associate with this DB instance.

If you do not specify a value for DBParameterGroupName, then Amazon RDS uses the DBParameterGroup of source DB instance for a same region read replica, or the default DBParameterGroup for the specified DB engine for a cross region read replica.

Currently, specifying a parameter group for this operation is only supported for Oracle DB instances.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

$sel:availabilityZone:CreateDBInstanceReadReplica', createDBInstanceReadReplica_availabilityZone - The Availability Zone (AZ) where the read replica will be created.

Default: A random, system-chosen Availability Zone in the endpoint's Amazon Web Services Region.

Example: us-east-1d

$sel:performanceInsightsRetentionPeriod:CreateDBInstanceReadReplica', createDBInstanceReadReplica_performanceInsightsRetentionPeriod - The amount of time, in days, to retain Performance Insights data. Valid values are 7 or 731 (2 years).

$sel:processorFeatures:CreateDBInstanceReadReplica', createDBInstanceReadReplica_processorFeatures - The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

$sel:tags:CreateDBInstanceReadReplica', createDBInstanceReadReplica_tags - Undocumented member.

$sel:dbInstanceClass:CreateDBInstanceReadReplica', createDBInstanceReadReplica_dbInstanceClass - The compute and memory capacity of the read replica, for example, db.m4.large. Not all DB instance classes are available in all Amazon Web Services Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

Default: Inherits from the source DB instance.

$sel:port:CreateDBInstanceReadReplica', createDBInstanceReadReplica_port - The port number that the DB instance uses for connections.

Default: Inherits from the source DB instance

Valid Values: 1150-65535

$sel:domainIAMRoleName:CreateDBInstanceReadReplica', createDBInstanceReadReplica_domainIAMRoleName - Specify the name of the IAM role to be used when making API calls to the Directory Service.

$sel:preSignedUrl:CreateDBInstanceReadReplica', createDBInstanceReadReplica_preSignedUrl - The URL that contains a Signature Version 4 signed request for the CreateDBInstanceReadReplica API action in the source Amazon Web Services Region that contains the source DB instance.

You must specify this parameter when you create an encrypted read replica from another Amazon Web Services Region by using the Amazon RDS API. Don't specify PreSignedUrl when you are creating an encrypted read replica in the same Amazon Web Services Region.

The presigned URL must be a valid request for the CreateDBInstanceReadReplica API action that can be executed in the source Amazon Web Services Region that contains the encrypted source DB instance. The presigned URL request must contain the following parameter values:

DestinationRegion - The Amazon Web Services Region that the encrypted read replica is created in. This Amazon Web Services Region is the same one where the CreateDBInstanceReadReplica action is called that contains this presigned URL.

For example, if you create an encrypted DB instance in the us-west-1 Amazon Web Services Region, from a source DB instance in the us-east-2 Amazon Web Services Region, then you call the CreateDBInstanceReadReplica action in the us-east-1 Amazon Web Services Region and provide a presigned URL that contains a call to the CreateDBInstanceReadReplica action in the us-west-2 Amazon Web Services Region. For this example, the DestinationRegion in the presigned URL must be set to the us-east-1 Amazon Web Services Region.

KmsKeyId - The Amazon Web Services KMS key identifier for the key to use to encrypt the read replica in the destination Amazon Web Services Region. This is the same identifier for both the CreateDBInstanceReadReplica action that is called in the destination Amazon Web Services Region, and the action contained in the presigned URL.

SourceDBInstanceIdentifier - The DB instance identifier for the encrypted DB instance to be replicated. This identifier must be in the Amazon Resource Name (ARN) format for the source Amazon Web Services Region. For example, if you are creating an encrypted read replica from a DB instance in the us-west-2 Amazon Web Services Region, then your SourceDBInstanceIdentifier looks like the following example: arn:aws:rds:us-west-2:123456789012:instance:mysql-instance1-20161115.

To learn how to generate a Signature Version 4 signed request, see Authenticating Requests: Using Query Parameters (Amazon Web Services Signature Version 4) and Signature Version 4 Signing Process.

If you are using an Amazon Web Services SDK tool or the CLI, you can specify SourceRegion (or --source-region for the CLI) instead of specifying PreSignedUrl manually. Specifying SourceRegion autogenerates a presigned URL that is a valid request for the operation that can be executed in the source Amazon Web Services Region.

SourceRegion isn't supported for SQL Server, because SQL Server on Amazon RDS doesn't support cross-region read replicas.

$sel:replicaMode:CreateDBInstanceReadReplica', createDBInstanceReadReplica_replicaMode - The open mode of the replica database: mounted or read-only.

This parameter is only supported for Oracle DB instances.

Mounted DB replicas are included in Oracle Enterprise Edition. The main use case for mounted replicas is cross-Region disaster recovery. The primary database doesn't use Active Data Guard to transmit information to the mounted replica. Because it doesn't accept user connections, a mounted replica can't serve a read-only workload.

You can create a combination of mounted and read-only DB replicas for the same primary DB instance. For more information, see Working with Oracle Read Replicas for Amazon RDS in the Amazon RDS User Guide.

$sel:copyTagsToSnapshot:CreateDBInstanceReadReplica', createDBInstanceReadReplica_copyTagsToSnapshot - A value that indicates whether to copy all tags from the read replica to snapshots of the read replica. By default, tags are not copied.

$sel:iops:CreateDBInstanceReadReplica', createDBInstanceReadReplica_iops - The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance.

$sel:autoMinorVersionUpgrade:CreateDBInstanceReadReplica', createDBInstanceReadReplica_autoMinorVersionUpgrade - A value that indicates whether minor engine upgrades are applied automatically to the read replica during the maintenance window.

Default: Inherits from the source DB instance

$sel:dbInstanceIdentifier:CreateDBInstanceReadReplica', createDBInstanceReadReplica_dbInstanceIdentifier - The DB instance identifier of the read replica. This identifier is the unique key that identifies a DB instance. This parameter is stored as a lowercase string.

$sel:sourceDBInstanceIdentifier:CreateDBInstanceReadReplica', createDBInstanceReadReplica_sourceDBInstanceIdentifier - The identifier of the DB instance that will act as the source for the read replica. Each DB instance can have up to five read replicas.

Constraints:

Must be the identifier of an existing MySQL, MariaDB, Oracle, PostgreSQL, or SQL Server DB instance.

Can specify a DB instance that is a MySQL read replica only if the source is running MySQL 5.6 or later.

For the limitations of Oracle read replicas, see Read Replica Limitations with Oracle in the Amazon RDS User Guide.

For the limitations of SQL Server read replicas, see Read Replica Limitations with Microsoft SQL Server in the Amazon RDS User Guide.

Can specify a PostgreSQL DB instance only if the source is running PostgreSQL 9.3.5 or later (9.4.7 and higher for cross-region replication).

The specified DB instance must have automatic backups enabled, that is, its backup retention period must be greater than 0.

If the source DB instance is in the same Amazon Web Services Region as the read replica, specify a valid DB instance identifier.

If the source DB instance is in a different Amazon Web Services Region from the read replica, specify a valid DB instance ARN. For more information, see Constructing an ARN for Amazon RDS in the Amazon RDS User Guide. This doesn't apply to SQL Server, which doesn't support cross-region replicas.

Request Lenses

createDBInstanceReadReplica_deletionProtection :: Lens' CreateDBInstanceReadReplica (Maybe Bool)

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

createDBInstanceReadReplica_enablePerformanceInsights :: Lens' CreateDBInstanceReadReplica (Maybe Bool)

A value that indicates whether to enable Performance Insights for the read replica.

For more information, see Using Amazon Performance Insights in the Amazon RDS User Guide.

createDBInstanceReadReplica_destinationRegion :: Lens' CreateDBInstanceReadReplica (Maybe Text)

Pseudo-parameter used when populating the PreSignedUrl of a cross-region CreateDBInstanceReadReplica request. To replicate from region SRC to region DST, send a request to region DST. In that request, pass a PreSignedUrl for region SRC with DestinationRegion set to region DST.

createDBInstanceReadReplica_maxAllocatedStorage :: Lens' CreateDBInstanceReadReplica (Maybe Int)

The upper limit in gibibytes (GiB) to which Amazon RDS can automatically scale the storage of the DB instance.

For more information about this setting, including limitations that apply to it, see Managing capacity automatically with Amazon RDS storage autoscaling in the Amazon RDS User Guide.

createDBInstanceReadReplica_enableIAMDatabaseAuthentication :: Lens' CreateDBInstanceReadReplica (Maybe Bool)

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information about IAM database authentication, see IAM Database Authentication for MySQL and PostgreSQL in the Amazon RDS User Guide.

createDBInstanceReadReplica_useDefaultProcessorFeatures :: Lens' CreateDBInstanceReadReplica (Maybe Bool)

A value that indicates whether the DB instance class of the DB instance uses its default processor features.

createDBInstanceReadReplica_enableCloudwatchLogsExports :: Lens' CreateDBInstanceReadReplica (Maybe [Text])

The list of logs that the new DB instance is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon RDS User Guide.

createDBInstanceReadReplica_storageType :: Lens' CreateDBInstanceReadReplica (Maybe Text)

Specifies the storage type to be associated with the read replica.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified, otherwise gp2

createDBInstanceReadReplica_optionGroupName :: Lens' CreateDBInstanceReadReplica (Maybe Text)

The option group the DB instance is associated with. If omitted, the option group associated with the source instance is used.

For SQL Server, you must use the option group associated with the source instance.

createDBInstanceReadReplica_monitoringInterval :: Lens' CreateDBInstanceReadReplica (Maybe Int)

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the read replica. To disable collecting Enhanced Monitoring metrics, specify 0. The default is 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

createDBInstanceReadReplica_domain :: Lens' CreateDBInstanceReadReplica (Maybe Text)

The Active Directory directory ID to create the DB instance in. Currently, only MySQL, Microsoft SQL Server, Oracle, and PostgreSQL DB instances can be created in an Active Directory Domain.

For more information, see Kerberos Authentication in the Amazon RDS User Guide.

createDBInstanceReadReplica_monitoringRoleArn :: Lens' CreateDBInstanceReadReplica (Maybe Text)

The ARN for the IAM role that permits RDS to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess. For information on creating a monitoring role, go to To create an IAM role for Amazon RDS Enhanced Monitoring in the Amazon RDS User Guide.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

createDBInstanceReadReplica_dbSubnetGroupName :: Lens' CreateDBInstanceReadReplica (Maybe Text)

Specifies a DB subnet group for the DB instance. The new DB instance is created in the VPC associated with the DB subnet group. If no DB subnet group is specified, then the new DB instance isn't created in a VPC.

Constraints:

Can only be specified if the source DB instance identifier specifies a DB instance in another Amazon Web Services Region.

If supplied, must match the name of an existing DBSubnetGroup.

The specified DB subnet group must be in the same Amazon Web Services Region in which the operation is running.

All read replicas in one Amazon Web Services Region that are created from the same source DB instance must either:>

Specify DB subnet groups from the same VPC. All these read replicas are created in the same VPC.

Not specify a DB subnet group. All these read replicas are created outside of any VPC.

Example: mySubnetgroup

createDBInstanceReadReplica_publiclyAccessible :: Lens' CreateDBInstanceReadReplica (Maybe Bool)

A value that indicates whether the DB instance is publicly accessible.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

For more information, see CreateDBInstance.

createDBInstanceReadReplica_multiAZ :: Lens' CreateDBInstanceReadReplica (Maybe Bool)

A value that indicates whether the read replica is in a Multi-AZ deployment.

You can create a read replica as a Multi-AZ DB instance. RDS creates a standby of your replica in another Availability Zone for failover support for the replica. Creating your read replica as a Multi-AZ DB instance is independent of whether the source database is a Multi-AZ DB instance.

createDBInstanceReadReplica_performanceInsightsKMSKeyId :: Lens' CreateDBInstanceReadReplica (Maybe Text)

The Amazon Web Services KMS key identifier for encryption of Performance Insights data.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

If you do not specify a value for PerformanceInsightsKMSKeyId, then Amazon RDS uses your default CMK. There is a default CMK for your Amazon Web Services account. Your Amazon Web Services account has a different default CMK for each Amazon Web Services Region.

createDBInstanceReadReplica_vpcSecurityGroupIds :: Lens' CreateDBInstanceReadReplica (Maybe [Text])

A list of EC2 VPC security groups to associate with the read replica.

Default: The default EC2 VPC security group for the DB subnet group's VPC.

createDBInstanceReadReplica_kmsKeyId :: Lens' CreateDBInstanceReadReplica (Maybe Text)

The Amazon Web Services KMS key identifier for an encrypted read replica.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS CMK.

If you create an encrypted read replica in the same Amazon Web Services Region as the source DB instance, then do not specify a value for this parameter. A read replica in the same Region is always encrypted with the same Amazon Web Services KMS CMK as the source DB instance.

If you create an encrypted read replica in a different Amazon Web Services Region, then you must specify a Amazon Web Services KMS key identifier for the destination Amazon Web Services Region. Amazon Web Services KMS CMKs are specific to the Amazon Web Services Region that they are created in, and you can't use CMKs from one Amazon Web Services Region in another Amazon Web Services Region.

You can't create an encrypted read replica from an unencrypted DB instance.

createDBInstanceReadReplica_dbParameterGroupName :: Lens' CreateDBInstanceReadReplica (Maybe Text)

The name of the DB parameter group to associate with this DB instance.

If you do not specify a value for DBParameterGroupName, then Amazon RDS uses the DBParameterGroup of source DB instance for a same region read replica, or the default DBParameterGroup for the specified DB engine for a cross region read replica.

Currently, specifying a parameter group for this operation is only supported for Oracle DB instances.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

createDBInstanceReadReplica_availabilityZone :: Lens' CreateDBInstanceReadReplica (Maybe Text)

The Availability Zone (AZ) where the read replica will be created.

Default: A random, system-chosen Availability Zone in the endpoint's Amazon Web Services Region.

Example: us-east-1d

createDBInstanceReadReplica_performanceInsightsRetentionPeriod :: Lens' CreateDBInstanceReadReplica (Maybe Int)

The amount of time, in days, to retain Performance Insights data. Valid values are 7 or 731 (2 years).

createDBInstanceReadReplica_processorFeatures :: Lens' CreateDBInstanceReadReplica (Maybe [ProcessorFeature])

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

createDBInstanceReadReplica_tags :: Lens' CreateDBInstanceReadReplica (Maybe [Tag])

Undocumented member.

createDBInstanceReadReplica_dbInstanceClass :: Lens' CreateDBInstanceReadReplica (Maybe Text)

The compute and memory capacity of the read replica, for example, db.m4.large. Not all DB instance classes are available in all Amazon Web Services Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

Default: Inherits from the source DB instance.

createDBInstanceReadReplica_port :: Lens' CreateDBInstanceReadReplica (Maybe Int)

The port number that the DB instance uses for connections.

Default: Inherits from the source DB instance

Valid Values: 1150-65535

createDBInstanceReadReplica_domainIAMRoleName :: Lens' CreateDBInstanceReadReplica (Maybe Text)

Specify the name of the IAM role to be used when making API calls to the Directory Service.

createDBInstanceReadReplica_preSignedUrl :: Lens' CreateDBInstanceReadReplica (Maybe Text)

The URL that contains a Signature Version 4 signed request for the CreateDBInstanceReadReplica API action in the source Amazon Web Services Region that contains the source DB instance.

You must specify this parameter when you create an encrypted read replica from another Amazon Web Services Region by using the Amazon RDS API. Don't specify PreSignedUrl when you are creating an encrypted read replica in the same Amazon Web Services Region.

The presigned URL must be a valid request for the CreateDBInstanceReadReplica API action that can be executed in the source Amazon Web Services Region that contains the encrypted source DB instance. The presigned URL request must contain the following parameter values:

DestinationRegion - The Amazon Web Services Region that the encrypted read replica is created in. This Amazon Web Services Region is the same one where the CreateDBInstanceReadReplica action is called that contains this presigned URL.

For example, if you create an encrypted DB instance in the us-west-1 Amazon Web Services Region, from a source DB instance in the us-east-2 Amazon Web Services Region, then you call the CreateDBInstanceReadReplica action in the us-east-1 Amazon Web Services Region and provide a presigned URL that contains a call to the CreateDBInstanceReadReplica action in the us-west-2 Amazon Web Services Region. For this example, the DestinationRegion in the presigned URL must be set to the us-east-1 Amazon Web Services Region.

KmsKeyId - The Amazon Web Services KMS key identifier for the key to use to encrypt the read replica in the destination Amazon Web Services Region. This is the same identifier for both the CreateDBInstanceReadReplica action that is called in the destination Amazon Web Services Region, and the action contained in the presigned URL.

SourceDBInstanceIdentifier - The DB instance identifier for the encrypted DB instance to be replicated. This identifier must be in the Amazon Resource Name (ARN) format for the source Amazon Web Services Region. For example, if you are creating an encrypted read replica from a DB instance in the us-west-2 Amazon Web Services Region, then your SourceDBInstanceIdentifier looks like the following example: arn:aws:rds:us-west-2:123456789012:instance:mysql-instance1-20161115.

To learn how to generate a Signature Version 4 signed request, see Authenticating Requests: Using Query Parameters (Amazon Web Services Signature Version 4) and Signature Version 4 Signing Process.

If you are using an Amazon Web Services SDK tool or the CLI, you can specify SourceRegion (or --source-region for the CLI) instead of specifying PreSignedUrl manually. Specifying SourceRegion autogenerates a presigned URL that is a valid request for the operation that can be executed in the source Amazon Web Services Region.

SourceRegion isn't supported for SQL Server, because SQL Server on Amazon RDS doesn't support cross-region read replicas.

createDBInstanceReadReplica_replicaMode :: Lens' CreateDBInstanceReadReplica (Maybe ReplicaMode)

The open mode of the replica database: mounted or read-only.

This parameter is only supported for Oracle DB instances.

Mounted DB replicas are included in Oracle Enterprise Edition. The main use case for mounted replicas is cross-Region disaster recovery. The primary database doesn't use Active Data Guard to transmit information to the mounted replica. Because it doesn't accept user connections, a mounted replica can't serve a read-only workload.

You can create a combination of mounted and read-only DB replicas for the same primary DB instance. For more information, see Working with Oracle Read Replicas for Amazon RDS in the Amazon RDS User Guide.

createDBInstanceReadReplica_copyTagsToSnapshot :: Lens' CreateDBInstanceReadReplica (Maybe Bool)

A value that indicates whether to copy all tags from the read replica to snapshots of the read replica. By default, tags are not copied.

createDBInstanceReadReplica_iops :: Lens' CreateDBInstanceReadReplica (Maybe Int)

The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance.

createDBInstanceReadReplica_autoMinorVersionUpgrade :: Lens' CreateDBInstanceReadReplica (Maybe Bool)

A value that indicates whether minor engine upgrades are applied automatically to the read replica during the maintenance window.

Default: Inherits from the source DB instance

createDBInstanceReadReplica_dbInstanceIdentifier :: Lens' CreateDBInstanceReadReplica Text

The DB instance identifier of the read replica. This identifier is the unique key that identifies a DB instance. This parameter is stored as a lowercase string.

createDBInstanceReadReplica_sourceDBInstanceIdentifier :: Lens' CreateDBInstanceReadReplica Text

The identifier of the DB instance that will act as the source for the read replica. Each DB instance can have up to five read replicas.

Constraints:

Must be the identifier of an existing MySQL, MariaDB, Oracle, PostgreSQL, or SQL Server DB instance.

Can specify a DB instance that is a MySQL read replica only if the source is running MySQL 5.6 or later.

For the limitations of Oracle read replicas, see Read Replica Limitations with Oracle in the Amazon RDS User Guide.

For the limitations of SQL Server read replicas, see Read Replica Limitations with Microsoft SQL Server in the Amazon RDS User Guide.

Can specify a PostgreSQL DB instance only if the source is running PostgreSQL 9.3.5 or later (9.4.7 and higher for cross-region replication).

The specified DB instance must have automatic backups enabled, that is, its backup retention period must be greater than 0.

If the source DB instance is in the same Amazon Web Services Region as the read replica, specify a valid DB instance identifier.

If the source DB instance is in a different Amazon Web Services Region from the read replica, specify a valid DB instance ARN. For more information, see Constructing an ARN for Amazon RDS in the Amazon RDS User Guide. This doesn't apply to SQL Server, which doesn't support cross-region replicas.

Destructuring the Response

data CreateDBInstanceReadReplicaResponse

See: newCreateDBInstanceReadReplicaResponse smart constructor.

Constructors

CreateDBInstanceReadReplicaResponse'

Fields

dbInstance :: Maybe DBInstance

httpStatus :: Int

The response's http status code.

newCreateDBInstanceReadReplicaResponse

Arguments

:: Int
$sel:httpStatus:CreateDBInstanceReadReplicaResponse'

-> CreateDBInstanceReadReplicaResponse

Create a value of CreateDBInstanceReadReplicaResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbInstance:CreateDBInstanceReadReplicaResponse', createDBInstanceReadReplicaResponse_dbInstance - Undocumented member.

$sel:httpStatus:CreateDBInstanceReadReplicaResponse', createDBInstanceReadReplicaResponse_httpStatus - The response's http status code.

Response Lenses

createDBInstanceReadReplicaResponse_dbInstance :: Lens' CreateDBInstanceReadReplicaResponse (Maybe DBInstance)

Undocumented member.

createDBInstanceReadReplicaResponse_httpStatus :: Lens' CreateDBInstanceReadReplicaResponse Int

The response's http status code.