Network.AWS.RDS.CreateDBProxy

Description

Creates a new DB proxy.

Creating a Request

data CreateDBProxy

See: newCreateDBProxy smart constructor.

Constructors

CreateDBProxy'

Fields

idleClientTimeout :: Maybe Int

The number of seconds that a connection to the proxy can be inactive before the proxy disconnects it. You can set this value higher or lower than the connection timeout limit for the associated database.

debugLogging :: Maybe Bool

Whether the proxy includes detailed information about SQL statements in its logs. This information helps you to debug issues involving SQL behavior or the performance and scalability of the proxy connections. The debug information includes the text of SQL statements that you submit through the proxy. Thus, only enable this setting when needed for debugging, and only when you have security measures in place to safeguard any sensitive information that appears in the logs.

vpcSecurityGroupIds :: Maybe [Text]

One or more VPC security group IDs to associate with the new proxy.

tags :: Maybe [Tag]

An optional set of key-value pairs to associate arbitrary data of your choosing with the proxy.

requireTLS :: Maybe Bool

A Boolean parameter that specifies whether Transport Layer Security (TLS) encryption is required for connections to the proxy. By enabling this setting, you can enforce encrypted TLS connections to the proxy.

dbProxyName :: Text

The identifier for the proxy. This name must be unique for all proxies owned by your Amazon Web Services account in the specified Amazon Web Services Region. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens; it can't end with a hyphen or contain two consecutive hyphens.

engineFamily :: EngineFamily

The kinds of databases that the proxy can connect to. This value determines which database network protocol the proxy recognizes when it interprets network traffic to and from the database. The engine family applies to MySQL and PostgreSQL for both RDS and Aurora.

auth :: [UserAuthConfig]

The authorization mechanism that the proxy uses.

roleArn :: Text

The Amazon Resource Name (ARN) of the IAM role that the proxy uses to access secrets in Amazon Web Services Secrets Manager.

vpcSubnetIds :: [Text]

One or more VPC subnet IDs to associate with the new proxy.

newCreateDBProxy

Arguments

:: Text
$sel:dbProxyName:CreateDBProxy'

-> EngineFamily
$sel:engineFamily:CreateDBProxy'

-> Text
$sel:roleArn:CreateDBProxy'

-> CreateDBProxy

Create a value of CreateDBProxy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:idleClientTimeout:CreateDBProxy', createDBProxy_idleClientTimeout - The number of seconds that a connection to the proxy can be inactive before the proxy disconnects it. You can set this value higher or lower than the connection timeout limit for the associated database.

$sel:debugLogging:CreateDBProxy', createDBProxy_debugLogging - Whether the proxy includes detailed information about SQL statements in its logs. This information helps you to debug issues involving SQL behavior or the performance and scalability of the proxy connections. The debug information includes the text of SQL statements that you submit through the proxy. Thus, only enable this setting when needed for debugging, and only when you have security measures in place to safeguard any sensitive information that appears in the logs.

$sel:vpcSecurityGroupIds:CreateDBProxy', createDBProxy_vpcSecurityGroupIds - One or more VPC security group IDs to associate with the new proxy.

$sel:tags:CreateDBProxy', createDBProxy_tags - An optional set of key-value pairs to associate arbitrary data of your choosing with the proxy.

$sel:requireTLS:CreateDBProxy', createDBProxy_requireTLS - A Boolean parameter that specifies whether Transport Layer Security (TLS) encryption is required for connections to the proxy. By enabling this setting, you can enforce encrypted TLS connections to the proxy.

$sel:dbProxyName:CreateDBProxy', createDBProxy_dbProxyName - The identifier for the proxy. This name must be unique for all proxies owned by your Amazon Web Services account in the specified Amazon Web Services Region. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens; it can't end with a hyphen or contain two consecutive hyphens.

$sel:engineFamily:CreateDBProxy', createDBProxy_engineFamily - The kinds of databases that the proxy can connect to. This value determines which database network protocol the proxy recognizes when it interprets network traffic to and from the database. The engine family applies to MySQL and PostgreSQL for both RDS and Aurora.

$sel:auth:CreateDBProxy', createDBProxy_auth - The authorization mechanism that the proxy uses.

$sel:roleArn:CreateDBProxy', createDBProxy_roleArn - The Amazon Resource Name (ARN) of the IAM role that the proxy uses to access secrets in Amazon Web Services Secrets Manager.

$sel:vpcSubnetIds:CreateDBProxy', createDBProxy_vpcSubnetIds - One or more VPC subnet IDs to associate with the new proxy.

Request Lenses

createDBProxy_idleClientTimeout :: Lens' CreateDBProxy (Maybe Int)

The number of seconds that a connection to the proxy can be inactive before the proxy disconnects it. You can set this value higher or lower than the connection timeout limit for the associated database.

createDBProxy_debugLogging :: Lens' CreateDBProxy (Maybe Bool)

Whether the proxy includes detailed information about SQL statements in its logs. This information helps you to debug issues involving SQL behavior or the performance and scalability of the proxy connections. The debug information includes the text of SQL statements that you submit through the proxy. Thus, only enable this setting when needed for debugging, and only when you have security measures in place to safeguard any sensitive information that appears in the logs.

createDBProxy_vpcSecurityGroupIds :: Lens' CreateDBProxy (Maybe [Text])

One or more VPC security group IDs to associate with the new proxy.

createDBProxy_tags :: Lens' CreateDBProxy (Maybe [Tag])

An optional set of key-value pairs to associate arbitrary data of your choosing with the proxy.

createDBProxy_requireTLS :: Lens' CreateDBProxy (Maybe Bool)

A Boolean parameter that specifies whether Transport Layer Security (TLS) encryption is required for connections to the proxy. By enabling this setting, you can enforce encrypted TLS connections to the proxy.

createDBProxy_dbProxyName :: Lens' CreateDBProxy Text

The identifier for the proxy. This name must be unique for all proxies owned by your Amazon Web Services account in the specified Amazon Web Services Region. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens; it can't end with a hyphen or contain two consecutive hyphens.

createDBProxy_engineFamily :: Lens' CreateDBProxy EngineFamily

The kinds of databases that the proxy can connect to. This value determines which database network protocol the proxy recognizes when it interprets network traffic to and from the database. The engine family applies to MySQL and PostgreSQL for both RDS and Aurora.

createDBProxy_auth :: Lens' CreateDBProxy [UserAuthConfig]

The authorization mechanism that the proxy uses.

createDBProxy_roleArn :: Lens' CreateDBProxy Text

The Amazon Resource Name (ARN) of the IAM role that the proxy uses to access secrets in Amazon Web Services Secrets Manager.

createDBProxy_vpcSubnetIds :: Lens' CreateDBProxy [Text]

One or more VPC subnet IDs to associate with the new proxy.

Destructuring the Response

data CreateDBProxyResponse

See: newCreateDBProxyResponse smart constructor.

Constructors

CreateDBProxyResponse'

Fields

dbProxy :: Maybe DBProxy

The DBProxy structure corresponding to the new proxy.

httpStatus :: Int

The response's http status code.

newCreateDBProxyResponse

Arguments

:: Int
$sel:httpStatus:CreateDBProxyResponse'

-> CreateDBProxyResponse

Create a value of CreateDBProxyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbProxy:CreateDBProxyResponse', createDBProxyResponse_dbProxy - The DBProxy structure corresponding to the new proxy.

$sel:httpStatus:CreateDBProxyResponse', createDBProxyResponse_httpStatus - The response's http status code.

Response Lenses

createDBProxyResponse_dbProxy :: Lens' CreateDBProxyResponse (Maybe DBProxy)

The DBProxy structure corresponding to the new proxy.

createDBProxyResponse_httpStatus :: Lens' CreateDBProxyResponse Int

The response's http status code.