Network.AWS.RDS.CreateDBProxyEndpoint

Description

Creates a DBProxyEndpoint. Only applies to proxies that are associated with Aurora DB clusters. You can use DB proxy endpoints to specify read/write or read-only access to the DB cluster. You can also use DB proxy endpoints to access a DB proxy through a different VPC than the proxy's default VPC.

Creating a Request

data CreateDBProxyEndpoint

See: newCreateDBProxyEndpoint smart constructor.

Constructors

CreateDBProxyEndpoint'

Fields

targetRole :: Maybe DBProxyEndpointTargetRole

A value that indicates whether the DB proxy endpoint can be used for read/write or read-only operations. The default is READ_WRITE.

vpcSecurityGroupIds :: Maybe [Text]

The VPC security group IDs for the DB proxy endpoint that you create. You can specify a different set of security group IDs than for the original DB proxy. The default is the default security group for the VPC.

tags :: Maybe [Tag]

dbProxyName :: Text

The name of the DB proxy associated with the DB proxy endpoint that you create.

dbProxyEndpointName :: Text

The name of the DB proxy endpoint to create.

vpcSubnetIds :: [Text]

The VPC subnet IDs for the DB proxy endpoint that you create. You can specify a different set of subnet IDs than for the original DB proxy.

newCreateDBProxyEndpoint

Arguments

:: Text
$sel:dbProxyName:CreateDBProxyEndpoint'

-> Text
$sel:dbProxyEndpointName:CreateDBProxyEndpoint'

-> CreateDBProxyEndpoint

Create a value of CreateDBProxyEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:targetRole:CreateDBProxyEndpoint', createDBProxyEndpoint_targetRole - A value that indicates whether the DB proxy endpoint can be used for read/write or read-only operations. The default is READ_WRITE.

$sel:vpcSecurityGroupIds:CreateDBProxyEndpoint', createDBProxyEndpoint_vpcSecurityGroupIds - The VPC security group IDs for the DB proxy endpoint that you create. You can specify a different set of security group IDs than for the original DB proxy. The default is the default security group for the VPC.

$sel:tags:CreateDBProxyEndpoint', createDBProxyEndpoint_tags - Undocumented member.

$sel:dbProxyName:CreateDBProxyEndpoint', createDBProxyEndpoint_dbProxyName - The name of the DB proxy associated with the DB proxy endpoint that you create.

$sel:dbProxyEndpointName:CreateDBProxyEndpoint', createDBProxyEndpoint_dbProxyEndpointName - The name of the DB proxy endpoint to create.

$sel:vpcSubnetIds:CreateDBProxyEndpoint', createDBProxyEndpoint_vpcSubnetIds - The VPC subnet IDs for the DB proxy endpoint that you create. You can specify a different set of subnet IDs than for the original DB proxy.

Request Lenses

createDBProxyEndpoint_targetRole :: Lens' CreateDBProxyEndpoint (Maybe DBProxyEndpointTargetRole)

A value that indicates whether the DB proxy endpoint can be used for read/write or read-only operations. The default is READ_WRITE.

createDBProxyEndpoint_vpcSecurityGroupIds :: Lens' CreateDBProxyEndpoint (Maybe [Text])

The VPC security group IDs for the DB proxy endpoint that you create. You can specify a different set of security group IDs than for the original DB proxy. The default is the default security group for the VPC.

createDBProxyEndpoint_tags :: Lens' CreateDBProxyEndpoint (Maybe [Tag])

Undocumented member.

createDBProxyEndpoint_dbProxyName :: Lens' CreateDBProxyEndpoint Text

The name of the DB proxy associated with the DB proxy endpoint that you create.

createDBProxyEndpoint_dbProxyEndpointName :: Lens' CreateDBProxyEndpoint Text

The name of the DB proxy endpoint to create.

createDBProxyEndpoint_vpcSubnetIds :: Lens' CreateDBProxyEndpoint [Text]

The VPC subnet IDs for the DB proxy endpoint that you create. You can specify a different set of subnet IDs than for the original DB proxy.

Destructuring the Response

data CreateDBProxyEndpointResponse

See: newCreateDBProxyEndpointResponse smart constructor.

Constructors

CreateDBProxyEndpointResponse'

Fields

dbProxyEndpoint :: Maybe DBProxyEndpoint

The DBProxyEndpoint object that is created by the API operation. The DB proxy endpoint that you create might provide capabilities such as read/write or read-only operations, or using a different VPC than the proxy's default VPC.

httpStatus :: Int

The response's http status code.

newCreateDBProxyEndpointResponse

Arguments

:: Int
$sel:httpStatus:CreateDBProxyEndpointResponse'

-> CreateDBProxyEndpointResponse

Create a value of CreateDBProxyEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbProxyEndpoint:CreateDBProxyEndpointResponse', createDBProxyEndpointResponse_dbProxyEndpoint - The DBProxyEndpoint object that is created by the API operation. The DB proxy endpoint that you create might provide capabilities such as read/write or read-only operations, or using a different VPC than the proxy's default VPC.

$sel:httpStatus:CreateDBProxyEndpointResponse', createDBProxyEndpointResponse_httpStatus - The response's http status code.

Response Lenses

createDBProxyEndpointResponse_dbProxyEndpoint :: Lens' CreateDBProxyEndpointResponse (Maybe DBProxyEndpoint)

The DBProxyEndpoint object that is created by the API operation. The DB proxy endpoint that you create might provide capabilities such as read/write or read-only operations, or using a different VPC than the proxy's default VPC.

createDBProxyEndpointResponse_httpStatus :: Lens' CreateDBProxyEndpointResponse Int

The response's http status code.