Network.AWS.RDS.CreateDBSnapshot

Description

Creates a snapshot of a DB instance. The source DB instance must be in the available or storage-optimization state.

Creating a Request

data CreateDBSnapshot

See: newCreateDBSnapshot smart constructor.

Constructors

CreateDBSnapshot'

Fields

tags :: Maybe [Tag]

dbSnapshotIdentifier :: Text

The identifier for the DB snapshot.

Constraints:

Can't be null, empty, or blank

Must contain from 1 to 255 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: my-snapshot-id

dbInstanceIdentifier :: Text

The identifier of the DB instance that you want to create the snapshot of.

Constraints:

Must match the identifier of an existing DBInstance.

newCreateDBSnapshot

Arguments

:: Text
$sel:dbSnapshotIdentifier:CreateDBSnapshot'

-> Text
$sel:dbInstanceIdentifier:CreateDBSnapshot'

-> CreateDBSnapshot

Create a value of CreateDBSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateDBSnapshot', createDBSnapshot_tags - Undocumented member.

$sel:dbSnapshotIdentifier:CreateDBSnapshot', createDBSnapshot_dbSnapshotIdentifier - The identifier for the DB snapshot.

Constraints:

Can't be null, empty, or blank

Must contain from 1 to 255 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: my-snapshot-id

$sel:dbInstanceIdentifier:CreateDBSnapshot', createDBSnapshot_dbInstanceIdentifier - The identifier of the DB instance that you want to create the snapshot of.

Constraints:

Must match the identifier of an existing DBInstance.

Request Lenses

createDBSnapshot_tags :: Lens' CreateDBSnapshot (Maybe [Tag])

Undocumented member.

createDBSnapshot_dbSnapshotIdentifier :: Lens' CreateDBSnapshot Text

The identifier for the DB snapshot.

Constraints:

Can't be null, empty, or blank

Must contain from 1 to 255 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: my-snapshot-id

createDBSnapshot_dbInstanceIdentifier :: Lens' CreateDBSnapshot Text

The identifier of the DB instance that you want to create the snapshot of.

Constraints:

Must match the identifier of an existing DBInstance.

Destructuring the Response

data CreateDBSnapshotResponse

See: newCreateDBSnapshotResponse smart constructor.

Constructors

CreateDBSnapshotResponse'

Fields

dbSnapshot :: Maybe DBSnapshot

httpStatus :: Int

The response's http status code.

newCreateDBSnapshotResponse

Arguments

:: Int
$sel:httpStatus:CreateDBSnapshotResponse'

-> CreateDBSnapshotResponse

Create a value of CreateDBSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbSnapshot:CreateDBSnapshotResponse', createDBSnapshotResponse_dbSnapshot - Undocumented member.

$sel:httpStatus:CreateDBSnapshotResponse', createDBSnapshotResponse_httpStatus - The response's http status code.

Response Lenses

createDBSnapshotResponse_dbSnapshot :: Lens' CreateDBSnapshotResponse (Maybe DBSnapshot)

Undocumented member.

createDBSnapshotResponse_httpStatus :: Lens' CreateDBSnapshotResponse Int

The response's http status code.