Network.AWS.RDS.CreateDBSubnetGroup

Description

Creates a new DB subnet group. DB subnet groups must contain at least one subnet in at least two AZs in the Amazon Web Services Region.

Creating a Request

data CreateDBSubnetGroup

See: newCreateDBSubnetGroup smart constructor.

Constructors

CreateDBSubnetGroup'

Fields

tags :: Maybe [Tag]

Tags to assign to the DB subnet group.

dbSubnetGroupName :: Text

The name for the DB subnet group. This value is stored as a lowercase string.

Constraints: Must contain no more than 255 letters, numbers, periods, underscores, spaces, or hyphens. Must not be default.

Example: mySubnetgroup

dbSubnetGroupDescription :: Text

The description for the DB subnet group.

subnetIds :: [Text]

The EC2 Subnet IDs for the DB subnet group.

newCreateDBSubnetGroup

Arguments

:: Text
$sel:dbSubnetGroupName:CreateDBSubnetGroup'

-> Text
$sel:dbSubnetGroupDescription:CreateDBSubnetGroup'

-> CreateDBSubnetGroup

Create a value of CreateDBSubnetGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateDBSubnetGroup', createDBSubnetGroup_tags - Tags to assign to the DB subnet group.

$sel:dbSubnetGroupName:CreateDBSubnetGroup', createDBSubnetGroup_dbSubnetGroupName - The name for the DB subnet group. This value is stored as a lowercase string.

Constraints: Must contain no more than 255 letters, numbers, periods, underscores, spaces, or hyphens. Must not be default.

Example: mySubnetgroup

$sel:dbSubnetGroupDescription:CreateDBSubnetGroup', createDBSubnetGroup_dbSubnetGroupDescription - The description for the DB subnet group.

$sel:subnetIds:CreateDBSubnetGroup', createDBSubnetGroup_subnetIds - The EC2 Subnet IDs for the DB subnet group.

Request Lenses

createDBSubnetGroup_tags :: Lens' CreateDBSubnetGroup (Maybe [Tag])

Tags to assign to the DB subnet group.

createDBSubnetGroup_dbSubnetGroupName :: Lens' CreateDBSubnetGroup Text

The name for the DB subnet group. This value is stored as a lowercase string.

Constraints: Must contain no more than 255 letters, numbers, periods, underscores, spaces, or hyphens. Must not be default.

Example: mySubnetgroup

createDBSubnetGroup_dbSubnetGroupDescription :: Lens' CreateDBSubnetGroup Text

The description for the DB subnet group.

createDBSubnetGroup_subnetIds :: Lens' CreateDBSubnetGroup [Text]

The EC2 Subnet IDs for the DB subnet group.

Destructuring the Response

data CreateDBSubnetGroupResponse

See: newCreateDBSubnetGroupResponse smart constructor.

Constructors

CreateDBSubnetGroupResponse'

Fields

dbSubnetGroup :: Maybe DBSubnetGroup

httpStatus :: Int

The response's http status code.

newCreateDBSubnetGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateDBSubnetGroupResponse'

-> CreateDBSubnetGroupResponse

Create a value of CreateDBSubnetGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbSubnetGroup:CreateDBSubnetGroupResponse', createDBSubnetGroupResponse_dbSubnetGroup - Undocumented member.

$sel:httpStatus:CreateDBSubnetGroupResponse', createDBSubnetGroupResponse_httpStatus - The response's http status code.

Response Lenses

createDBSubnetGroupResponse_dbSubnetGroup :: Lens' CreateDBSubnetGroupResponse (Maybe DBSubnetGroup)

Undocumented member.

createDBSubnetGroupResponse_httpStatus :: Lens' CreateDBSubnetGroupResponse Int

The response's http status code.