Network.AWS.RDS.CreateEventSubscription

Description

Creates an RDS event notification subscription. This action requires a topic Amazon Resource Name (ARN) created by either the RDS console, the SNS console, or the SNS API. To obtain an ARN with SNS, you must create a topic in Amazon SNS and subscribe to the topic. The ARN is displayed in the SNS console.

You can specify the type of source (SourceType) that you want to be notified of and provide a list of RDS sources (SourceIds) that triggers the events. You can also provide a list of event categories (EventCategories) for events that you want to be notified of. For example, you can specify SourceType = db-instance, SourceIds = mydbinstance1, mydbinstance2 and EventCategories = Availability, Backup.

If you specify both the SourceType and SourceIds, such as SourceType = db-instance and SourceIdentifier = myDBInstance1, you are notified of all the db-instance events for the specified source. If you specify a SourceType but do not specify a SourceIdentifier, you receive notice of the events for that source type for all your RDS sources. If you don't specify either the SourceType or the SourceIdentifier, you are notified of events generated from all RDS sources belonging to your customer account.

RDS event notification is only available for unencrypted SNS topics. If you specify an encrypted SNS topic, event notifications aren't sent for the topic.

Creating a Request

data CreateEventSubscription

See: newCreateEventSubscription smart constructor.

Constructors

CreateEventSubscription'

Fields

enabled :: Maybe Bool

A value that indicates whether to activate the subscription. If the event notification subscription isn't activated, the subscription is created but not active.

sourceType :: Maybe Text

The type of source that is generating the events. For example, if you want to be notified of events generated by a DB instance, you set this parameter to db-instance. If this value isn't specified, all events are returned.

Valid values: db-instance | db-cluster | db-parameter-group | db-security-group | db-snapshot | db-cluster-snapshot

eventCategories :: Maybe [Text]

A list of event categories for a particular source type (SourceType) that you want to subscribe to. You can see a list of the categories for a given source type in Events in the Amazon RDS User Guide or by using the DescribeEventCategories operation.

sourceIds :: Maybe [Text]

The list of identifiers of the event sources for which events are returned. If not specified, then all sources are included in the response. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens. It can't end with a hyphen or contain two consecutive hyphens.

Constraints:

If SourceIds are supplied, SourceType must also be provided.

If the source type is a DB instance, a DBInstanceIdentifier value must be supplied.

If the source type is a DB cluster, a DBClusterIdentifier value must be supplied.

If the source type is a DB parameter group, a DBParameterGroupName value must be supplied.

If the source type is a DB security group, a DBSecurityGroupName value must be supplied.

If the source type is a DB snapshot, a DBSnapshotIdentifier value must be supplied.

If the source type is a DB cluster snapshot, a DBClusterSnapshotIdentifier value must be supplied.

tags :: Maybe [Tag]

subscriptionName :: Text

The name of the subscription.

Constraints: The name must be less than 255 characters.

snsTopicArn :: Text

The Amazon Resource Name (ARN) of the SNS topic created for event notification. The ARN is created by Amazon SNS when you create a topic and subscribe to it.

newCreateEventSubscription

Arguments

:: Text
$sel:subscriptionName:CreateEventSubscription'

-> Text
$sel:snsTopicArn:CreateEventSubscription'

-> CreateEventSubscription

Create a value of CreateEventSubscription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enabled:CreateEventSubscription', createEventSubscription_enabled - A value that indicates whether to activate the subscription. If the event notification subscription isn't activated, the subscription is created but not active.

$sel:sourceType:CreateEventSubscription', createEventSubscription_sourceType - The type of source that is generating the events. For example, if you want to be notified of events generated by a DB instance, you set this parameter to db-instance. If this value isn't specified, all events are returned.

Valid values: db-instance | db-cluster | db-parameter-group | db-security-group | db-snapshot | db-cluster-snapshot

$sel:eventCategories:CreateEventSubscription', createEventSubscription_eventCategories - A list of event categories for a particular source type (SourceType) that you want to subscribe to. You can see a list of the categories for a given source type in Events in the Amazon RDS User Guide or by using the DescribeEventCategories operation.

$sel:sourceIds:CreateEventSubscription', createEventSubscription_sourceIds - The list of identifiers of the event sources for which events are returned. If not specified, then all sources are included in the response. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens. It can't end with a hyphen or contain two consecutive hyphens.

Constraints:

If SourceIds are supplied, SourceType must also be provided.

If the source type is a DB instance, a DBInstanceIdentifier value must be supplied.

If the source type is a DB cluster, a DBClusterIdentifier value must be supplied.

If the source type is a DB parameter group, a DBParameterGroupName value must be supplied.

If the source type is a DB security group, a DBSecurityGroupName value must be supplied.

If the source type is a DB snapshot, a DBSnapshotIdentifier value must be supplied.

If the source type is a DB cluster snapshot, a DBClusterSnapshotIdentifier value must be supplied.

$sel:tags:CreateEventSubscription', createEventSubscription_tags - Undocumented member.

$sel:subscriptionName:CreateEventSubscription', createEventSubscription_subscriptionName - The name of the subscription.

Constraints: The name must be less than 255 characters.

$sel:snsTopicArn:CreateEventSubscription', createEventSubscription_snsTopicArn - The Amazon Resource Name (ARN) of the SNS topic created for event notification. The ARN is created by Amazon SNS when you create a topic and subscribe to it.

Request Lenses

createEventSubscription_enabled :: Lens' CreateEventSubscription (Maybe Bool)

A value that indicates whether to activate the subscription. If the event notification subscription isn't activated, the subscription is created but not active.

createEventSubscription_sourceType :: Lens' CreateEventSubscription (Maybe Text)

The type of source that is generating the events. For example, if you want to be notified of events generated by a DB instance, you set this parameter to db-instance. If this value isn't specified, all events are returned.

Valid values: db-instance | db-cluster | db-parameter-group | db-security-group | db-snapshot | db-cluster-snapshot

createEventSubscription_eventCategories :: Lens' CreateEventSubscription (Maybe [Text])

A list of event categories for a particular source type (SourceType) that you want to subscribe to. You can see a list of the categories for a given source type in Events in the Amazon RDS User Guide or by using the DescribeEventCategories operation.

createEventSubscription_sourceIds :: Lens' CreateEventSubscription (Maybe [Text])

The list of identifiers of the event sources for which events are returned. If not specified, then all sources are included in the response. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens. It can't end with a hyphen or contain two consecutive hyphens.

Constraints:

If SourceIds are supplied, SourceType must also be provided.

If the source type is a DB instance, a DBInstanceIdentifier value must be supplied.

If the source type is a DB cluster, a DBClusterIdentifier value must be supplied.

If the source type is a DB parameter group, a DBParameterGroupName value must be supplied.

If the source type is a DB security group, a DBSecurityGroupName value must be supplied.

If the source type is a DB snapshot, a DBSnapshotIdentifier value must be supplied.

If the source type is a DB cluster snapshot, a DBClusterSnapshotIdentifier value must be supplied.

createEventSubscription_tags :: Lens' CreateEventSubscription (Maybe [Tag])

Undocumented member.

createEventSubscription_subscriptionName :: Lens' CreateEventSubscription Text

The name of the subscription.

Constraints: The name must be less than 255 characters.

createEventSubscription_snsTopicArn :: Lens' CreateEventSubscription Text

The Amazon Resource Name (ARN) of the SNS topic created for event notification. The ARN is created by Amazon SNS when you create a topic and subscribe to it.

Destructuring the Response

data CreateEventSubscriptionResponse

See: newCreateEventSubscriptionResponse smart constructor.

Constructors

CreateEventSubscriptionResponse'

Fields

eventSubscription :: Maybe EventSubscription

httpStatus :: Int

The response's http status code.

newCreateEventSubscriptionResponse

Arguments

:: Int
$sel:httpStatus:CreateEventSubscriptionResponse'

-> CreateEventSubscriptionResponse

Create a value of CreateEventSubscriptionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventSubscription:CreateEventSubscriptionResponse', createEventSubscriptionResponse_eventSubscription - Undocumented member.

$sel:httpStatus:CreateEventSubscriptionResponse', createEventSubscriptionResponse_httpStatus - The response's http status code.

Response Lenses

createEventSubscriptionResponse_eventSubscription :: Lens' CreateEventSubscriptionResponse (Maybe EventSubscription)

Undocumented member.

createEventSubscriptionResponse_httpStatus :: Lens' CreateEventSubscriptionResponse Int

The response's http status code.