Network.AWS.RDS.DeleteDBCluster

Description

The DeleteDBCluster action deletes a previously provisioned DB cluster. When you delete a DB cluster, all automated backups for that DB cluster are deleted and can't be recovered. Manual DB cluster snapshots of the specified DB cluster are not deleted.

For more information on Amazon Aurora, see What Is Amazon Aurora? in the Amazon Aurora User Guide.

This action only applies to Aurora DB clusters.

Creating a Request

data DeleteDBCluster

See: newDeleteDBCluster smart constructor.

Constructors

DeleteDBCluster'

Fields

skipFinalSnapshot :: Maybe Bool

A value that indicates whether to skip the creation of a final DB cluster snapshot before the DB cluster is deleted. If skip is specified, no DB cluster snapshot is created. If skip isn't specified, a DB cluster snapshot is created before the DB cluster is deleted. By default, skip isn't specified, and the DB cluster snapshot is created. By default, this parameter is disabled.

You must specify a FinalDBSnapshotIdentifier parameter if SkipFinalSnapshot is disabled.

finalDBSnapshotIdentifier :: Maybe Text

The DB cluster snapshot identifier of the new DB cluster snapshot created when SkipFinalSnapshot is disabled.

Specifying this parameter and also skipping the creation of a final DB cluster snapshot with the SkipFinalShapshot parameter results in an error.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

dbClusterIdentifier :: Text

The DB cluster identifier for the DB cluster to be deleted. This parameter isn't case-sensitive.

Constraints:

Must match an existing DBClusterIdentifier.

newDeleteDBCluster

Arguments

:: Text
$sel:dbClusterIdentifier:DeleteDBCluster'

-> DeleteDBCluster

Create a value of DeleteDBCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:skipFinalSnapshot:DeleteDBCluster', deleteDBCluster_skipFinalSnapshot - A value that indicates whether to skip the creation of a final DB cluster snapshot before the DB cluster is deleted. If skip is specified, no DB cluster snapshot is created. If skip isn't specified, a DB cluster snapshot is created before the DB cluster is deleted. By default, skip isn't specified, and the DB cluster snapshot is created. By default, this parameter is disabled.

You must specify a FinalDBSnapshotIdentifier parameter if SkipFinalSnapshot is disabled.

$sel:finalDBSnapshotIdentifier:DeleteDBCluster', deleteDBCluster_finalDBSnapshotIdentifier - The DB cluster snapshot identifier of the new DB cluster snapshot created when SkipFinalSnapshot is disabled.

Specifying this parameter and also skipping the creation of a final DB cluster snapshot with the SkipFinalShapshot parameter results in an error.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

$sel:dbClusterIdentifier:DeleteDBCluster', deleteDBCluster_dbClusterIdentifier - The DB cluster identifier for the DB cluster to be deleted. This parameter isn't case-sensitive.

Constraints:

Must match an existing DBClusterIdentifier.

Request Lenses

deleteDBCluster_skipFinalSnapshot :: Lens' DeleteDBCluster (Maybe Bool)

A value that indicates whether to skip the creation of a final DB cluster snapshot before the DB cluster is deleted. If skip is specified, no DB cluster snapshot is created. If skip isn't specified, a DB cluster snapshot is created before the DB cluster is deleted. By default, skip isn't specified, and the DB cluster snapshot is created. By default, this parameter is disabled.

You must specify a FinalDBSnapshotIdentifier parameter if SkipFinalSnapshot is disabled.

deleteDBCluster_finalDBSnapshotIdentifier :: Lens' DeleteDBCluster (Maybe Text)

The DB cluster snapshot identifier of the new DB cluster snapshot created when SkipFinalSnapshot is disabled.

Specifying this parameter and also skipping the creation of a final DB cluster snapshot with the SkipFinalShapshot parameter results in an error.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

deleteDBCluster_dbClusterIdentifier :: Lens' DeleteDBCluster Text

The DB cluster identifier for the DB cluster to be deleted. This parameter isn't case-sensitive.

Constraints:

Must match an existing DBClusterIdentifier.

Destructuring the Response

data DeleteDBClusterResponse

See: newDeleteDBClusterResponse smart constructor.

Constructors

DeleteDBClusterResponse'

Fields

dbCluster :: Maybe DBCluster

httpStatus :: Int

The response's http status code.

newDeleteDBClusterResponse

Arguments

:: Int
$sel:httpStatus:DeleteDBClusterResponse'

-> DeleteDBClusterResponse

Create a value of DeleteDBClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbCluster:DeleteDBClusterResponse', deleteDBClusterResponse_dbCluster - Undocumented member.

$sel:httpStatus:DeleteDBClusterResponse', deleteDBClusterResponse_httpStatus - The response's http status code.

Response Lenses

deleteDBClusterResponse_dbCluster :: Lens' DeleteDBClusterResponse (Maybe DBCluster)

Undocumented member.

deleteDBClusterResponse_httpStatus :: Lens' DeleteDBClusterResponse Int

The response's http status code.