Network.AWS.RDS.DeleteDBClusterSnapshot

Description

Deletes a DB cluster snapshot. If the snapshot is being copied, the copy operation is terminated.

The DB cluster snapshot must be in the available state to be deleted.

For more information on Amazon Aurora, see What Is Amazon Aurora? in the Amazon Aurora User Guide.

This action only applies to Aurora DB clusters.

Creating a Request

data DeleteDBClusterSnapshot

See: newDeleteDBClusterSnapshot smart constructor.

Constructors

DeleteDBClusterSnapshot'

Fields

dbClusterSnapshotIdentifier :: Text

The identifier of the DB cluster snapshot to delete.

Constraints: Must be the name of an existing DB cluster snapshot in the available state.

newDeleteDBClusterSnapshot

Arguments

:: Text
$sel:dbClusterSnapshotIdentifier:DeleteDBClusterSnapshot'

-> DeleteDBClusterSnapshot

Create a value of DeleteDBClusterSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbClusterSnapshotIdentifier:DeleteDBClusterSnapshot', deleteDBClusterSnapshot_dbClusterSnapshotIdentifier - The identifier of the DB cluster snapshot to delete.

Constraints: Must be the name of an existing DB cluster snapshot in the available state.

Request Lenses

deleteDBClusterSnapshot_dbClusterSnapshotIdentifier :: Lens' DeleteDBClusterSnapshot Text

The identifier of the DB cluster snapshot to delete.

Constraints: Must be the name of an existing DB cluster snapshot in the available state.

Destructuring the Response

data DeleteDBClusterSnapshotResponse

See: newDeleteDBClusterSnapshotResponse smart constructor.

Constructors

DeleteDBClusterSnapshotResponse'

Fields

dbClusterSnapshot :: Maybe DBClusterSnapshot

httpStatus :: Int

The response's http status code.

newDeleteDBClusterSnapshotResponse

Arguments

:: Int
$sel:httpStatus:DeleteDBClusterSnapshotResponse'

-> DeleteDBClusterSnapshotResponse

Create a value of DeleteDBClusterSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbClusterSnapshot:DeleteDBClusterSnapshotResponse', deleteDBClusterSnapshotResponse_dbClusterSnapshot - Undocumented member.

$sel:httpStatus:DeleteDBClusterSnapshotResponse', deleteDBClusterSnapshotResponse_httpStatus - The response's http status code.

Response Lenses

deleteDBClusterSnapshotResponse_dbClusterSnapshot :: Lens' DeleteDBClusterSnapshotResponse (Maybe DBClusterSnapshot)

Undocumented member.

deleteDBClusterSnapshotResponse_httpStatus :: Lens' DeleteDBClusterSnapshotResponse Int

The response's http status code.