Network.AWS.RDS.DeleteDBSecurityGroup

Description

Deletes a DB security group.

The specified DB security group must not be associated with any DB instances.

Creating a Request

data DeleteDBSecurityGroup

See: newDeleteDBSecurityGroup smart constructor.

Constructors

DeleteDBSecurityGroup'

Fields

dbSecurityGroupName :: Text

The name of the DB security group to delete.

You can't delete the default DB security group.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Must not be "Default"

newDeleteDBSecurityGroup

Arguments

:: Text
$sel:dbSecurityGroupName:DeleteDBSecurityGroup'

-> DeleteDBSecurityGroup

Create a value of DeleteDBSecurityGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbSecurityGroupName:DeleteDBSecurityGroup', deleteDBSecurityGroup_dbSecurityGroupName - The name of the DB security group to delete.

You can't delete the default DB security group.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Must not be "Default"

Request Lenses

deleteDBSecurityGroup_dbSecurityGroupName :: Lens' DeleteDBSecurityGroup Text

The name of the DB security group to delete.

You can't delete the default DB security group.

Constraints:

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Must not be "Default"

Destructuring the Response

data DeleteDBSecurityGroupResponse

See: newDeleteDBSecurityGroupResponse smart constructor.

Constructors

DeleteDBSecurityGroupResponse'

newDeleteDBSecurityGroupResponse :: DeleteDBSecurityGroupResponse

Create a value of DeleteDBSecurityGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.