Network.AWS.RDS.DeregisterDBProxyTargets

Description

Remove the association between one or more DBProxyTarget data structures and a DBProxyTargetGroup.

Creating a Request

data DeregisterDBProxyTargets

See: newDeregisterDBProxyTargets smart constructor.

Constructors

DeregisterDBProxyTargets'

Fields

dbClusterIdentifiers :: Maybe [Text]

One or more DB cluster identifiers.

targetGroupName :: Maybe Text

The identifier of the DBProxyTargetGroup.

dbInstanceIdentifiers :: Maybe [Text]

One or more DB instance identifiers.

dbProxyName :: Text

The identifier of the DBProxy that is associated with the DBProxyTargetGroup.

newDeregisterDBProxyTargets

Arguments

:: Text
$sel:dbProxyName:DeregisterDBProxyTargets'

-> DeregisterDBProxyTargets

Create a value of DeregisterDBProxyTargets with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbClusterIdentifiers:DeregisterDBProxyTargets', deregisterDBProxyTargets_dbClusterIdentifiers - One or more DB cluster identifiers.

$sel:targetGroupName:DeregisterDBProxyTargets', deregisterDBProxyTargets_targetGroupName - The identifier of the DBProxyTargetGroup.

$sel:dbInstanceIdentifiers:DeregisterDBProxyTargets', deregisterDBProxyTargets_dbInstanceIdentifiers - One or more DB instance identifiers.

$sel:dbProxyName:DeregisterDBProxyTargets', deregisterDBProxyTargets_dbProxyName - The identifier of the DBProxy that is associated with the DBProxyTargetGroup.

Request Lenses

deregisterDBProxyTargets_dbClusterIdentifiers :: Lens' DeregisterDBProxyTargets (Maybe [Text])

One or more DB cluster identifiers.

deregisterDBProxyTargets_targetGroupName :: Lens' DeregisterDBProxyTargets (Maybe Text)

The identifier of the DBProxyTargetGroup.

deregisterDBProxyTargets_dbInstanceIdentifiers :: Lens' DeregisterDBProxyTargets (Maybe [Text])

One or more DB instance identifiers.

deregisterDBProxyTargets_dbProxyName :: Lens' DeregisterDBProxyTargets Text

The identifier of the DBProxy that is associated with the DBProxyTargetGroup.

Destructuring the Response

data DeregisterDBProxyTargetsResponse

See: newDeregisterDBProxyTargetsResponse smart constructor.

Constructors

DeregisterDBProxyTargetsResponse'

Fields

httpStatus :: Int

The response's http status code.

newDeregisterDBProxyTargetsResponse

Arguments

:: Int
$sel:httpStatus:DeregisterDBProxyTargetsResponse'

-> DeregisterDBProxyTargetsResponse

Create a value of DeregisterDBProxyTargetsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DeregisterDBProxyTargetsResponse', deregisterDBProxyTargetsResponse_httpStatus - The response's http status code.

Response Lenses

deregisterDBProxyTargetsResponse_httpStatus :: Lens' DeregisterDBProxyTargetsResponse Int

The response's http status code.