Network.AWS.RDS.DescribeAccountAttributes

Description

Lists all of the attributes for a customer account. The attributes include Amazon RDS quotas for the account, such as the number of DB instances allowed. The description for a quota includes the quota name, current usage toward that quota, and the quota's maximum value.

This command doesn't take any parameters.

Creating a Request

data DescribeAccountAttributes

See: newDescribeAccountAttributes smart constructor.

Constructors

DescribeAccountAttributes'

newDescribeAccountAttributes :: DescribeAccountAttributes

Create a value of DescribeAccountAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

Destructuring the Response

data DescribeAccountAttributesResponse

Data returned by the DescribeAccountAttributes action.

See: newDescribeAccountAttributesResponse smart constructor.

Constructors

DescribeAccountAttributesResponse'

Fields

accountQuotas :: Maybe [AccountQuota]

A list of AccountQuota objects. Within this list, each quota has a name, a count of usage toward the quota maximum, and a maximum value for the quota.

httpStatus :: Int

The response's http status code.

newDescribeAccountAttributesResponse

Arguments

:: Int
$sel:httpStatus:DescribeAccountAttributesResponse'

-> DescribeAccountAttributesResponse

Create a value of DescribeAccountAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:accountQuotas:DescribeAccountAttributesResponse', describeAccountAttributesResponse_accountQuotas - A list of AccountQuota objects. Within this list, each quota has a name, a count of usage toward the quota maximum, and a maximum value for the quota.

$sel:httpStatus:DescribeAccountAttributesResponse', describeAccountAttributesResponse_httpStatus - The response's http status code.

Response Lenses

describeAccountAttributesResponse_accountQuotas :: Lens' DescribeAccountAttributesResponse (Maybe [AccountQuota])

A list of AccountQuota objects. Within this list, each quota has a name, a count of usage toward the quota maximum, and a maximum value for the quota.

describeAccountAttributesResponse_httpStatus :: Lens' DescribeAccountAttributesResponse Int

The response's http status code.