Network.AWS.RDS.DescribeDBClusterBacktracks

Description

Returns information about backtracks for a DB cluster.

For more information on Amazon Aurora, see What Is Amazon Aurora? in the Amazon Aurora User Guide.

This action only applies to Aurora MySQL DB clusters.

This operation returns paginated results.

Creating a Request

data DescribeDBClusterBacktracks

See: newDescribeDBClusterBacktracks smart constructor.

Constructors

DescribeDBClusterBacktracks'

Fields

backtrackIdentifier :: Maybe Text

If specified, this value is the backtrack identifier of the backtrack to be described.

Constraints:

Must contain a valid universally unique identifier (UUID). For more information about UUIDs, see A Universally Unique Identifier (UUID) URN Namespace.

Example: 123e4567-e89b-12d3-a456-426655440000

filters :: Maybe [Filter]

A filter that specifies one or more DB clusters to describe. Supported filters include the following:

db-cluster-backtrack-id - Accepts backtrack identifiers. The results list includes information about only the backtracks identified by these identifiers.

db-cluster-backtrack-status - Accepts any of the following backtrack status values:

applying

completed

failed

pending

The results list includes information about only the backtracks identified by these values.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

marker :: Maybe Text

An optional pagination token provided by a previous DescribeDBClusterBacktracks request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dbClusterIdentifier :: Text

The DB cluster identifier of the DB cluster to be described. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: my-cluster1

newDescribeDBClusterBacktracks

Arguments

:: Text
$sel:dbClusterIdentifier:DescribeDBClusterBacktracks'

-> DescribeDBClusterBacktracks

Create a value of DescribeDBClusterBacktracks with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backtrackIdentifier:DescribeDBClusterBacktracks', describeDBClusterBacktracks_backtrackIdentifier - If specified, this value is the backtrack identifier of the backtrack to be described.

Constraints:

Must contain a valid universally unique identifier (UUID). For more information about UUIDs, see A Universally Unique Identifier (UUID) URN Namespace.

Example: 123e4567-e89b-12d3-a456-426655440000

$sel:filters:DescribeDBClusterBacktracks', describeDBClusterBacktracks_filters - A filter that specifies one or more DB clusters to describe. Supported filters include the following:

db-cluster-backtrack-id - Accepts backtrack identifiers. The results list includes information about only the backtracks identified by these identifiers.

db-cluster-backtrack-status - Accepts any of the following backtrack status values:

applying

completed

failed

pending

The results list includes information about only the backtracks identified by these values.

$sel:maxRecords:DescribeDBClusterBacktracks', describeDBClusterBacktracks_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

$sel:marker:DescribeDBClusterBacktracks', describeDBClusterBacktracks_marker - An optional pagination token provided by a previous DescribeDBClusterBacktracks request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:dbClusterIdentifier:DescribeDBClusterBacktracks', describeDBClusterBacktracks_dbClusterIdentifier - The DB cluster identifier of the DB cluster to be described. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: my-cluster1

Request Lenses

describeDBClusterBacktracks_backtrackIdentifier :: Lens' DescribeDBClusterBacktracks (Maybe Text)

If specified, this value is the backtrack identifier of the backtrack to be described.

Constraints:

Must contain a valid universally unique identifier (UUID). For more information about UUIDs, see A Universally Unique Identifier (UUID) URN Namespace.

Example: 123e4567-e89b-12d3-a456-426655440000

describeDBClusterBacktracks_filters :: Lens' DescribeDBClusterBacktracks (Maybe [Filter])

A filter that specifies one or more DB clusters to describe. Supported filters include the following:

db-cluster-backtrack-id - Accepts backtrack identifiers. The results list includes information about only the backtracks identified by these identifiers.

db-cluster-backtrack-status - Accepts any of the following backtrack status values:

applying

completed

failed

pending

The results list includes information about only the backtracks identified by these values.

describeDBClusterBacktracks_maxRecords :: Lens' DescribeDBClusterBacktracks (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

describeDBClusterBacktracks_marker :: Lens' DescribeDBClusterBacktracks (Maybe Text)

An optional pagination token provided by a previous DescribeDBClusterBacktracks request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeDBClusterBacktracks_dbClusterIdentifier :: Lens' DescribeDBClusterBacktracks Text

The DB cluster identifier of the DB cluster to be described. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: my-cluster1

Destructuring the Response

data DescribeDBClusterBacktracksResponse

Contains the result of a successful invocation of the DescribeDBClusterBacktracks action.

See: newDescribeDBClusterBacktracksResponse smart constructor.

Constructors

DescribeDBClusterBacktracksResponse'

Fields

dbClusterBacktracks :: Maybe [DBClusterBacktrack]

Contains a list of backtracks for the user.

marker :: Maybe Text

A pagination token that can be used in a later DescribeDBClusterBacktracks request.

httpStatus :: Int

The response's http status code.

newDescribeDBClusterBacktracksResponse

Arguments

:: Int
$sel:httpStatus:DescribeDBClusterBacktracksResponse'

-> DescribeDBClusterBacktracksResponse

Create a value of DescribeDBClusterBacktracksResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbClusterBacktracks:DescribeDBClusterBacktracksResponse', describeDBClusterBacktracksResponse_dbClusterBacktracks - Contains a list of backtracks for the user.

$sel:marker:DescribeDBClusterBacktracks', describeDBClusterBacktracksResponse_marker - A pagination token that can be used in a later DescribeDBClusterBacktracks request.

$sel:httpStatus:DescribeDBClusterBacktracksResponse', describeDBClusterBacktracksResponse_httpStatus - The response's http status code.

Response Lenses

describeDBClusterBacktracksResponse_dbClusterBacktracks :: Lens' DescribeDBClusterBacktracksResponse (Maybe [DBClusterBacktrack])

Contains a list of backtracks for the user.

describeDBClusterBacktracksResponse_marker :: Lens' DescribeDBClusterBacktracksResponse (Maybe Text)

A pagination token that can be used in a later DescribeDBClusterBacktracks request.

describeDBClusterBacktracksResponse_httpStatus :: Lens' DescribeDBClusterBacktracksResponse Int

The response's http status code.