Network.AWS.RDS.DescribeDBClusterEndpoints

Description

Returns information about endpoints for an Amazon Aurora DB cluster.

This action only applies to Aurora DB clusters.

This operation returns paginated results.

Creating a Request

data DescribeDBClusterEndpoints

See: newDescribeDBClusterEndpoints smart constructor.

Constructors

DescribeDBClusterEndpoints'

Fields

dbClusterEndpointIdentifier :: Maybe Text

The identifier of the endpoint to describe. This parameter is stored as a lowercase string.

dbClusterIdentifier :: Maybe Text

The DB cluster identifier of the DB cluster associated with the endpoint. This parameter is stored as a lowercase string.

filters :: Maybe [Filter]

A set of name-value pairs that define which endpoints to include in the output. The filters are specified as name-value pairs, in the format Name=endpoint_type,Values=endpoint_type1,endpoint_type2,.... Name can be one of: db-cluster-endpoint-type, db-cluster-endpoint-custom-type, db-cluster-endpoint-id, db-cluster-endpoint-status. Values for the db-cluster-endpoint-type filter can be one or more of: reader, writer, custom. Values for the db-cluster-endpoint-custom-type filter can be one or more of: reader, any. Values for the db-cluster-endpoint-status filter can be one or more of: available, creating, deleting, inactive, modifying.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

marker :: Maybe Text

An optional pagination token provided by a previous DescribeDBClusterEndpoints request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

newDescribeDBClusterEndpoints :: DescribeDBClusterEndpoints

Create a value of DescribeDBClusterEndpoints with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbClusterEndpointIdentifier:DescribeDBClusterEndpoints', describeDBClusterEndpoints_dbClusterEndpointIdentifier - The identifier of the endpoint to describe. This parameter is stored as a lowercase string.

$sel:dbClusterIdentifier:DescribeDBClusterEndpoints', describeDBClusterEndpoints_dbClusterIdentifier - The DB cluster identifier of the DB cluster associated with the endpoint. This parameter is stored as a lowercase string.

$sel:filters:DescribeDBClusterEndpoints', describeDBClusterEndpoints_filters - A set of name-value pairs that define which endpoints to include in the output. The filters are specified as name-value pairs, in the format Name=endpoint_type,Values=endpoint_type1,endpoint_type2,.... Name can be one of: db-cluster-endpoint-type, db-cluster-endpoint-custom-type, db-cluster-endpoint-id, db-cluster-endpoint-status. Values for the db-cluster-endpoint-type filter can be one or more of: reader, writer, custom. Values for the db-cluster-endpoint-custom-type filter can be one or more of: reader, any. Values for the db-cluster-endpoint-status filter can be one or more of: available, creating, deleting, inactive, modifying.

$sel:maxRecords:DescribeDBClusterEndpoints', describeDBClusterEndpoints_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

$sel:marker:DescribeDBClusterEndpoints', describeDBClusterEndpoints_marker - An optional pagination token provided by a previous DescribeDBClusterEndpoints request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Request Lenses

describeDBClusterEndpoints_dbClusterEndpointIdentifier :: Lens' DescribeDBClusterEndpoints (Maybe Text)

The identifier of the endpoint to describe. This parameter is stored as a lowercase string.

describeDBClusterEndpoints_dbClusterIdentifier :: Lens' DescribeDBClusterEndpoints (Maybe Text)

The DB cluster identifier of the DB cluster associated with the endpoint. This parameter is stored as a lowercase string.

describeDBClusterEndpoints_filters :: Lens' DescribeDBClusterEndpoints (Maybe [Filter])

A set of name-value pairs that define which endpoints to include in the output. The filters are specified as name-value pairs, in the format Name=endpoint_type,Values=endpoint_type1,endpoint_type2,.... Name can be one of: db-cluster-endpoint-type, db-cluster-endpoint-custom-type, db-cluster-endpoint-id, db-cluster-endpoint-status. Values for the db-cluster-endpoint-type filter can be one or more of: reader, writer, custom. Values for the db-cluster-endpoint-custom-type filter can be one or more of: reader, any. Values for the db-cluster-endpoint-status filter can be one or more of: available, creating, deleting, inactive, modifying.

describeDBClusterEndpoints_maxRecords :: Lens' DescribeDBClusterEndpoints (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

describeDBClusterEndpoints_marker :: Lens' DescribeDBClusterEndpoints (Maybe Text)

An optional pagination token provided by a previous DescribeDBClusterEndpoints request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Destructuring the Response

data DescribeDBClusterEndpointsResponse

See: newDescribeDBClusterEndpointsResponse smart constructor.

Constructors

DescribeDBClusterEndpointsResponse'

Fields

dbClusterEndpoints :: Maybe [DBClusterEndpoint]

Contains the details of the endpoints associated with the cluster and matching any filter conditions.

marker :: Maybe Text

An optional pagination token provided by a previous DescribeDBClusterEndpoints request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

httpStatus :: Int

The response's http status code.

newDescribeDBClusterEndpointsResponse

Arguments

:: Int
$sel:httpStatus:DescribeDBClusterEndpointsResponse'

-> DescribeDBClusterEndpointsResponse

Create a value of DescribeDBClusterEndpointsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbClusterEndpoints:DescribeDBClusterEndpointsResponse', describeDBClusterEndpointsResponse_dbClusterEndpoints - Contains the details of the endpoints associated with the cluster and matching any filter conditions.

$sel:marker:DescribeDBClusterEndpoints', describeDBClusterEndpointsResponse_marker - An optional pagination token provided by a previous DescribeDBClusterEndpoints request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:httpStatus:DescribeDBClusterEndpointsResponse', describeDBClusterEndpointsResponse_httpStatus - The response's http status code.

Response Lenses

describeDBClusterEndpointsResponse_dbClusterEndpoints :: Lens' DescribeDBClusterEndpointsResponse (Maybe [DBClusterEndpoint])

Contains the details of the endpoints associated with the cluster and matching any filter conditions.

describeDBClusterEndpointsResponse_marker :: Lens' DescribeDBClusterEndpointsResponse (Maybe Text)

An optional pagination token provided by a previous DescribeDBClusterEndpoints request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeDBClusterEndpointsResponse_httpStatus :: Lens' DescribeDBClusterEndpointsResponse Int

The response's http status code.