Network.AWS.RDS.DescribeDBClusterSnapshots

Description

Returns information about DB cluster snapshots. This API action supports pagination.

For more information on Amazon Aurora, see What Is Amazon Aurora? in the Amazon Aurora User Guide.

This action only applies to Aurora DB clusters.

This operation returns paginated results.

Creating a Request

data DescribeDBClusterSnapshots

See: newDescribeDBClusterSnapshots smart constructor.

Constructors

DescribeDBClusterSnapshots'

Fields

dbClusterIdentifier :: Maybe Text

The ID of the DB cluster to retrieve the list of DB cluster snapshots for. This parameter can't be used in conjunction with the DBClusterSnapshotIdentifier parameter. This parameter isn't case-sensitive.

Constraints:

If supplied, must match the identifier of an existing DBCluster.

includeShared :: Maybe Bool

A value that indicates whether to include shared manual DB cluster snapshots from other Amazon Web Services accounts that this Amazon Web Services account has been given permission to copy or restore. By default, these snapshots are not included.

You can give an Amazon Web Services account permission to restore a manual DB cluster snapshot from another Amazon Web Services account by the ModifyDBClusterSnapshotAttribute API action.

dbClusterSnapshotIdentifier :: Maybe Text

A specific DB cluster snapshot identifier to describe. This parameter can't be used in conjunction with the DBClusterIdentifier parameter. This value is stored as a lowercase string.

Constraints:

If supplied, must match the identifier of an existing DBClusterSnapshot.

If this identifier is for an automated snapshot, the SnapshotType parameter must also be specified.

filters :: Maybe [Filter]

A filter that specifies one or more DB cluster snapshots to describe.

Supported filters:

db-cluster-id - Accepts DB cluster identifiers and DB cluster Amazon Resource Names (ARNs).

db-cluster-snapshot-id - Accepts DB cluster snapshot identifiers.

snapshot-type - Accepts types of DB cluster snapshots.

engine - Accepts names of database engines.

snapshotType :: Maybe Text

The type of DB cluster snapshots to be returned. You can specify one of the following values:

automated - Return all DB cluster snapshots that have been automatically taken by Amazon RDS for my Amazon Web Services account.

manual - Return all DB cluster snapshots that have been taken by my Amazon Web Services account.

shared - Return all manual DB cluster snapshots that have been shared to my Amazon Web Services account.

public - Return all DB cluster snapshots that have been marked as public.

If you don't specify a SnapshotType value, then both automated and manual DB cluster snapshots are returned. You can include shared DB cluster snapshots with these results by enabling the IncludeShared parameter. You can include public DB cluster snapshots with these results by enabling the IncludePublic parameter.

The IncludeShared and IncludePublic parameters don't apply for SnapshotType values of manual or automated. The IncludePublic parameter doesn't apply when SnapshotType is set to shared. The IncludeShared parameter doesn't apply when SnapshotType is set to public.

marker :: Maybe Text

An optional pagination token provided by a previous DescribeDBClusterSnapshots request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

includePublic :: Maybe Bool

A value that indicates whether to include manual DB cluster snapshots that are public and can be copied or restored by any Amazon Web Services account. By default, the public snapshots are not included.

You can share a manual DB cluster snapshot as public by using the ModifyDBClusterSnapshotAttribute API action.

newDescribeDBClusterSnapshots :: DescribeDBClusterSnapshots

Create a value of DescribeDBClusterSnapshots with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbClusterIdentifier:DescribeDBClusterSnapshots', describeDBClusterSnapshots_dbClusterIdentifier - The ID of the DB cluster to retrieve the list of DB cluster snapshots for. This parameter can't be used in conjunction with the DBClusterSnapshotIdentifier parameter. This parameter isn't case-sensitive.

Constraints:

If supplied, must match the identifier of an existing DBCluster.

$sel:includeShared:DescribeDBClusterSnapshots', describeDBClusterSnapshots_includeShared - A value that indicates whether to include shared manual DB cluster snapshots from other Amazon Web Services accounts that this Amazon Web Services account has been given permission to copy or restore. By default, these snapshots are not included.

You can give an Amazon Web Services account permission to restore a manual DB cluster snapshot from another Amazon Web Services account by the ModifyDBClusterSnapshotAttribute API action.

$sel:dbClusterSnapshotIdentifier:DescribeDBClusterSnapshots', describeDBClusterSnapshots_dbClusterSnapshotIdentifier - A specific DB cluster snapshot identifier to describe. This parameter can't be used in conjunction with the DBClusterIdentifier parameter. This value is stored as a lowercase string.

Constraints:

If supplied, must match the identifier of an existing DBClusterSnapshot.

If this identifier is for an automated snapshot, the SnapshotType parameter must also be specified.

$sel:filters:DescribeDBClusterSnapshots', describeDBClusterSnapshots_filters - A filter that specifies one or more DB cluster snapshots to describe.

Supported filters:

db-cluster-id - Accepts DB cluster identifiers and DB cluster Amazon Resource Names (ARNs).

db-cluster-snapshot-id - Accepts DB cluster snapshot identifiers.

snapshot-type - Accepts types of DB cluster snapshots.

engine - Accepts names of database engines.

$sel:snapshotType:DescribeDBClusterSnapshots', describeDBClusterSnapshots_snapshotType - The type of DB cluster snapshots to be returned. You can specify one of the following values:

automated - Return all DB cluster snapshots that have been automatically taken by Amazon RDS for my Amazon Web Services account.

manual - Return all DB cluster snapshots that have been taken by my Amazon Web Services account.

shared - Return all manual DB cluster snapshots that have been shared to my Amazon Web Services account.

public - Return all DB cluster snapshots that have been marked as public.

If you don't specify a SnapshotType value, then both automated and manual DB cluster snapshots are returned. You can include shared DB cluster snapshots with these results by enabling the IncludeShared parameter. You can include public DB cluster snapshots with these results by enabling the IncludePublic parameter.

The IncludeShared and IncludePublic parameters don't apply for SnapshotType values of manual or automated. The IncludePublic parameter doesn't apply when SnapshotType is set to shared. The IncludeShared parameter doesn't apply when SnapshotType is set to public.

$sel:marker:DescribeDBClusterSnapshots', describeDBClusterSnapshots_marker - An optional pagination token provided by a previous DescribeDBClusterSnapshots request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:maxRecords:DescribeDBClusterSnapshots', describeDBClusterSnapshots_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

$sel:includePublic:DescribeDBClusterSnapshots', describeDBClusterSnapshots_includePublic - A value that indicates whether to include manual DB cluster snapshots that are public and can be copied or restored by any Amazon Web Services account. By default, the public snapshots are not included.

You can share a manual DB cluster snapshot as public by using the ModifyDBClusterSnapshotAttribute API action.

Request Lenses

describeDBClusterSnapshots_dbClusterIdentifier :: Lens' DescribeDBClusterSnapshots (Maybe Text)

The ID of the DB cluster to retrieve the list of DB cluster snapshots for. This parameter can't be used in conjunction with the DBClusterSnapshotIdentifier parameter. This parameter isn't case-sensitive.

Constraints:

If supplied, must match the identifier of an existing DBCluster.

describeDBClusterSnapshots_includeShared :: Lens' DescribeDBClusterSnapshots (Maybe Bool)

A value that indicates whether to include shared manual DB cluster snapshots from other Amazon Web Services accounts that this Amazon Web Services account has been given permission to copy or restore. By default, these snapshots are not included.

You can give an Amazon Web Services account permission to restore a manual DB cluster snapshot from another Amazon Web Services account by the ModifyDBClusterSnapshotAttribute API action.

describeDBClusterSnapshots_dbClusterSnapshotIdentifier :: Lens' DescribeDBClusterSnapshots (Maybe Text)

A specific DB cluster snapshot identifier to describe. This parameter can't be used in conjunction with the DBClusterIdentifier parameter. This value is stored as a lowercase string.

Constraints:

If supplied, must match the identifier of an existing DBClusterSnapshot.

If this identifier is for an automated snapshot, the SnapshotType parameter must also be specified.

describeDBClusterSnapshots_filters :: Lens' DescribeDBClusterSnapshots (Maybe [Filter])

A filter that specifies one or more DB cluster snapshots to describe.

Supported filters:

db-cluster-id - Accepts DB cluster identifiers and DB cluster Amazon Resource Names (ARNs).

db-cluster-snapshot-id - Accepts DB cluster snapshot identifiers.

snapshot-type - Accepts types of DB cluster snapshots.

engine - Accepts names of database engines.

describeDBClusterSnapshots_snapshotType :: Lens' DescribeDBClusterSnapshots (Maybe Text)

The type of DB cluster snapshots to be returned. You can specify one of the following values:

automated - Return all DB cluster snapshots that have been automatically taken by Amazon RDS for my Amazon Web Services account.

manual - Return all DB cluster snapshots that have been taken by my Amazon Web Services account.

shared - Return all manual DB cluster snapshots that have been shared to my Amazon Web Services account.

public - Return all DB cluster snapshots that have been marked as public.

If you don't specify a SnapshotType value, then both automated and manual DB cluster snapshots are returned. You can include shared DB cluster snapshots with these results by enabling the IncludeShared parameter. You can include public DB cluster snapshots with these results by enabling the IncludePublic parameter.

The IncludeShared and IncludePublic parameters don't apply for SnapshotType values of manual or automated. The IncludePublic parameter doesn't apply when SnapshotType is set to shared. The IncludeShared parameter doesn't apply when SnapshotType is set to public.

describeDBClusterSnapshots_marker :: Lens' DescribeDBClusterSnapshots (Maybe Text)

An optional pagination token provided by a previous DescribeDBClusterSnapshots request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeDBClusterSnapshots_maxRecords :: Lens' DescribeDBClusterSnapshots (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

describeDBClusterSnapshots_includePublic :: Lens' DescribeDBClusterSnapshots (Maybe Bool)

A value that indicates whether to include manual DB cluster snapshots that are public and can be copied or restored by any Amazon Web Services account. By default, the public snapshots are not included.

You can share a manual DB cluster snapshot as public by using the ModifyDBClusterSnapshotAttribute API action.

Destructuring the Response

data DescribeDBClusterSnapshotsResponse

Provides a list of DB cluster snapshots for the user as the result of a call to the DescribeDBClusterSnapshots action.

See: newDescribeDBClusterSnapshotsResponse smart constructor.

Constructors

DescribeDBClusterSnapshotsResponse'

Fields

marker :: Maybe Text

An optional pagination token provided by a previous DescribeDBClusterSnapshots request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dbClusterSnapshots :: Maybe [DBClusterSnapshot]

Provides a list of DB cluster snapshots for the user.

httpStatus :: Int

The response's http status code.

newDescribeDBClusterSnapshotsResponse

Arguments

:: Int
$sel:httpStatus:DescribeDBClusterSnapshotsResponse'

-> DescribeDBClusterSnapshotsResponse

Create a value of DescribeDBClusterSnapshotsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:DescribeDBClusterSnapshots', describeDBClusterSnapshotsResponse_marker - An optional pagination token provided by a previous DescribeDBClusterSnapshots request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:dbClusterSnapshots:DescribeDBClusterSnapshotsResponse', describeDBClusterSnapshotsResponse_dbClusterSnapshots - Provides a list of DB cluster snapshots for the user.

$sel:httpStatus:DescribeDBClusterSnapshotsResponse', describeDBClusterSnapshotsResponse_httpStatus - The response's http status code.

Response Lenses

describeDBClusterSnapshotsResponse_marker :: Lens' DescribeDBClusterSnapshotsResponse (Maybe Text)

An optional pagination token provided by a previous DescribeDBClusterSnapshots request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeDBClusterSnapshotsResponse_dbClusterSnapshots :: Lens' DescribeDBClusterSnapshotsResponse (Maybe [DBClusterSnapshot])

Provides a list of DB cluster snapshots for the user.

describeDBClusterSnapshotsResponse_httpStatus :: Lens' DescribeDBClusterSnapshotsResponse Int

The response's http status code.