Network.AWS.RDS.DescribeDBClusters

Description

Returns information about provisioned Aurora DB clusters. This API supports pagination.

For more information on Amazon Aurora, see What Is Amazon Aurora? in the Amazon Aurora User Guide.

This operation can also return information for Amazon Neptune DB instances and Amazon DocumentDB instances.

This operation returns paginated results.

Creating a Request

data DescribeDBClusters

See: newDescribeDBClusters smart constructor.

Constructors

DescribeDBClusters'

Fields

dbClusterIdentifier :: Maybe Text

The user-supplied DB cluster identifier. If this parameter is specified, information from only the specific DB cluster is returned. This parameter isn't case-sensitive.

Constraints:

If supplied, must match an existing DBClusterIdentifier.

includeShared :: Maybe Bool

Optional Boolean parameter that specifies whether the output includes information about clusters shared from other Amazon Web Services accounts.

filters :: Maybe [Filter]

A filter that specifies one or more DB clusters to describe.

Supported filters:

clone-group-id - Accepts clone group identifiers. The results list will only include information about the DB clusters associated with these clone groups.

db-cluster-id - Accepts DB cluster identifiers and DB cluster Amazon Resource Names (ARNs). The results list will only include information about the DB clusters identified by these ARNs.

domain - Accepts Active Directory directory IDs. The results list will only include information about the DB clusters associated with these domains.

engine - Accepts engine names. The results list will only include information about the DB clusters for these engines.

marker :: Maybe Text

An optional pagination token provided by a previous DescribeDBClusters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

newDescribeDBClusters :: DescribeDBClusters

Create a value of DescribeDBClusters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbClusterIdentifier:DescribeDBClusters', describeDBClusters_dbClusterIdentifier - The user-supplied DB cluster identifier. If this parameter is specified, information from only the specific DB cluster is returned. This parameter isn't case-sensitive.

Constraints:

If supplied, must match an existing DBClusterIdentifier.

$sel:includeShared:DescribeDBClusters', describeDBClusters_includeShared - Optional Boolean parameter that specifies whether the output includes information about clusters shared from other Amazon Web Services accounts.

$sel:filters:DescribeDBClusters', describeDBClusters_filters - A filter that specifies one or more DB clusters to describe.

Supported filters:

clone-group-id - Accepts clone group identifiers. The results list will only include information about the DB clusters associated with these clone groups.

db-cluster-id - Accepts DB cluster identifiers and DB cluster Amazon Resource Names (ARNs). The results list will only include information about the DB clusters identified by these ARNs.

domain - Accepts Active Directory directory IDs. The results list will only include information about the DB clusters associated with these domains.

engine - Accepts engine names. The results list will only include information about the DB clusters for these engines.

$sel:marker:DescribeDBClusters', describeDBClusters_marker - An optional pagination token provided by a previous DescribeDBClusters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:maxRecords:DescribeDBClusters', describeDBClusters_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

Request Lenses

describeDBClusters_dbClusterIdentifier :: Lens' DescribeDBClusters (Maybe Text)

The user-supplied DB cluster identifier. If this parameter is specified, information from only the specific DB cluster is returned. This parameter isn't case-sensitive.

Constraints:

If supplied, must match an existing DBClusterIdentifier.

describeDBClusters_includeShared :: Lens' DescribeDBClusters (Maybe Bool)

Optional Boolean parameter that specifies whether the output includes information about clusters shared from other Amazon Web Services accounts.

describeDBClusters_filters :: Lens' DescribeDBClusters (Maybe [Filter])

A filter that specifies one or more DB clusters to describe.

Supported filters:

clone-group-id - Accepts clone group identifiers. The results list will only include information about the DB clusters associated with these clone groups.

db-cluster-id - Accepts DB cluster identifiers and DB cluster Amazon Resource Names (ARNs). The results list will only include information about the DB clusters identified by these ARNs.

domain - Accepts Active Directory directory IDs. The results list will only include information about the DB clusters associated with these domains.

engine - Accepts engine names. The results list will only include information about the DB clusters for these engines.

describeDBClusters_marker :: Lens' DescribeDBClusters (Maybe Text)

An optional pagination token provided by a previous DescribeDBClusters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeDBClusters_maxRecords :: Lens' DescribeDBClusters (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

Destructuring the Response

data DescribeDBClustersResponse

Contains the result of a successful invocation of the DescribeDBClusters action.

See: newDescribeDBClustersResponse smart constructor.

Constructors

DescribeDBClustersResponse'

Fields

dbClusters :: Maybe [DBCluster]

Contains a list of DB clusters for the user.

marker :: Maybe Text

A pagination token that can be used in a later DescribeDBClusters request.

httpStatus :: Int

The response's http status code.

newDescribeDBClustersResponse

Arguments

:: Int
$sel:httpStatus:DescribeDBClustersResponse'

-> DescribeDBClustersResponse

Create a value of DescribeDBClustersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbClusters:DescribeDBClustersResponse', describeDBClustersResponse_dbClusters - Contains a list of DB clusters for the user.

$sel:marker:DescribeDBClusters', describeDBClustersResponse_marker - A pagination token that can be used in a later DescribeDBClusters request.

$sel:httpStatus:DescribeDBClustersResponse', describeDBClustersResponse_httpStatus - The response's http status code.

Response Lenses

describeDBClustersResponse_dbClusters :: Lens' DescribeDBClustersResponse (Maybe [DBCluster])

Contains a list of DB clusters for the user.

describeDBClustersResponse_marker :: Lens' DescribeDBClustersResponse (Maybe Text)

A pagination token that can be used in a later DescribeDBClusters request.

describeDBClustersResponse_httpStatus :: Lens' DescribeDBClustersResponse Int

The response's http status code.