Network.AWS.RDS.DescribeDBInstanceAutomatedBackups

Description

Displays backups for both current and deleted instances. For example, use this operation to find details about automated backups for previously deleted instances. Current instances with retention periods greater than zero (0) are returned for both the DescribeDBInstanceAutomatedBackups and DescribeDBInstances operations.

All parameters are optional.

This operation returns paginated results.

Creating a Request

data DescribeDBInstanceAutomatedBackups

Parameter input for DescribeDBInstanceAutomatedBackups.

See: newDescribeDBInstanceAutomatedBackups smart constructor.

Constructors

DescribeDBInstanceAutomatedBackups'

Fields

dbInstanceAutomatedBackupsArn :: Maybe Text

The Amazon Resource Name (ARN) of the replicated automated backups, for example, arn:aws:rds:us-east-1:123456789012:auto-backup:ab-L2IJCEXJP7XQ7HOJ4SIEXAMPLE.

dbiResourceId :: Maybe Text

The resource ID of the DB instance that is the source of the automated backup. This parameter isn't case-sensitive.

dbInstanceIdentifier :: Maybe Text

(Optional) The user-supplied instance identifier. If this parameter is specified, it must match the identifier of an existing DB instance. It returns information from the specific DB instance' automated backup. This parameter isn't case-sensitive.

filters :: Maybe [Filter]

A filter that specifies which resources to return based on status.

Supported filters are the following:

status

active - automated backups for current instances

retained - automated backups for deleted instances and after backup replication is stopped

creating - automated backups that are waiting for the first automated snapshot to be available

db-instance-id - Accepts DB instance identifiers and Amazon Resource Names (ARNs). The results list includes only information about the DB instance automated backups identified by these ARNs.

dbi-resource-id - Accepts DB resource identifiers and Amazon Resource Names (ARNs). The results list includes only information about the DB instance resources identified by these ARNs.

Returns all resources by default. The status for each resource is specified in the response.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

marker :: Maybe Text

The pagination token provided in the previous request. If this parameter is specified the response includes only records beyond the marker, up to MaxRecords.

newDescribeDBInstanceAutomatedBackups :: DescribeDBInstanceAutomatedBackups

Create a value of DescribeDBInstanceAutomatedBackups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbInstanceAutomatedBackupsArn:DescribeDBInstanceAutomatedBackups', describeDBInstanceAutomatedBackups_dbInstanceAutomatedBackupsArn - The Amazon Resource Name (ARN) of the replicated automated backups, for example, arn:aws:rds:us-east-1:123456789012:auto-backup:ab-L2IJCEXJP7XQ7HOJ4SIEXAMPLE.

$sel:dbiResourceId:DescribeDBInstanceAutomatedBackups', describeDBInstanceAutomatedBackups_dbiResourceId - The resource ID of the DB instance that is the source of the automated backup. This parameter isn't case-sensitive.

$sel:dbInstanceIdentifier:DescribeDBInstanceAutomatedBackups', describeDBInstanceAutomatedBackups_dbInstanceIdentifier - (Optional) The user-supplied instance identifier. If this parameter is specified, it must match the identifier of an existing DB instance. It returns information from the specific DB instance' automated backup. This parameter isn't case-sensitive.

$sel:filters:DescribeDBInstanceAutomatedBackups', describeDBInstanceAutomatedBackups_filters - A filter that specifies which resources to return based on status.

Supported filters are the following:

status

active - automated backups for current instances

retained - automated backups for deleted instances and after backup replication is stopped

creating - automated backups that are waiting for the first automated snapshot to be available

db-instance-id - Accepts DB instance identifiers and Amazon Resource Names (ARNs). The results list includes only information about the DB instance automated backups identified by these ARNs.

dbi-resource-id - Accepts DB resource identifiers and Amazon Resource Names (ARNs). The results list includes only information about the DB instance resources identified by these ARNs.

Returns all resources by default. The status for each resource is specified in the response.

$sel:maxRecords:DescribeDBInstanceAutomatedBackups', describeDBInstanceAutomatedBackups_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

$sel:marker:DescribeDBInstanceAutomatedBackups', describeDBInstanceAutomatedBackups_marker - The pagination token provided in the previous request. If this parameter is specified the response includes only records beyond the marker, up to MaxRecords.

Request Lenses

describeDBInstanceAutomatedBackups_dbInstanceAutomatedBackupsArn :: Lens' DescribeDBInstanceAutomatedBackups (Maybe Text)

The Amazon Resource Name (ARN) of the replicated automated backups, for example, arn:aws:rds:us-east-1:123456789012:auto-backup:ab-L2IJCEXJP7XQ7HOJ4SIEXAMPLE.

describeDBInstanceAutomatedBackups_dbiResourceId :: Lens' DescribeDBInstanceAutomatedBackups (Maybe Text)

The resource ID of the DB instance that is the source of the automated backup. This parameter isn't case-sensitive.

describeDBInstanceAutomatedBackups_dbInstanceIdentifier :: Lens' DescribeDBInstanceAutomatedBackups (Maybe Text)

(Optional) The user-supplied instance identifier. If this parameter is specified, it must match the identifier of an existing DB instance. It returns information from the specific DB instance' automated backup. This parameter isn't case-sensitive.

describeDBInstanceAutomatedBackups_filters :: Lens' DescribeDBInstanceAutomatedBackups (Maybe [Filter])

A filter that specifies which resources to return based on status.

Supported filters are the following:

status

active - automated backups for current instances

retained - automated backups for deleted instances and after backup replication is stopped

creating - automated backups that are waiting for the first automated snapshot to be available

db-instance-id - Accepts DB instance identifiers and Amazon Resource Names (ARNs). The results list includes only information about the DB instance automated backups identified by these ARNs.

dbi-resource-id - Accepts DB resource identifiers and Amazon Resource Names (ARNs). The results list includes only information about the DB instance resources identified by these ARNs.

Returns all resources by default. The status for each resource is specified in the response.

describeDBInstanceAutomatedBackups_maxRecords :: Lens' DescribeDBInstanceAutomatedBackups (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

describeDBInstanceAutomatedBackups_marker :: Lens' DescribeDBInstanceAutomatedBackups (Maybe Text)

The pagination token provided in the previous request. If this parameter is specified the response includes only records beyond the marker, up to MaxRecords.

Destructuring the Response

data DescribeDBInstanceAutomatedBackupsResponse

Contains the result of a successful invocation of the DescribeDBInstanceAutomatedBackups action.

See: newDescribeDBInstanceAutomatedBackupsResponse smart constructor.

Constructors

DescribeDBInstanceAutomatedBackupsResponse'

Fields

dbInstanceAutomatedBackups :: Maybe [DBInstanceAutomatedBackup]

A list of DBInstanceAutomatedBackup instances.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

httpStatus :: Int

The response's http status code.

newDescribeDBInstanceAutomatedBackupsResponse

Arguments

:: Int
$sel:httpStatus:DescribeDBInstanceAutomatedBackupsResponse'

-> DescribeDBInstanceAutomatedBackupsResponse

Create a value of DescribeDBInstanceAutomatedBackupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbInstanceAutomatedBackups:DescribeDBInstanceAutomatedBackupsResponse', describeDBInstanceAutomatedBackupsResponse_dbInstanceAutomatedBackups - A list of DBInstanceAutomatedBackup instances.

$sel:marker:DescribeDBInstanceAutomatedBackups', describeDBInstanceAutomatedBackupsResponse_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

$sel:httpStatus:DescribeDBInstanceAutomatedBackupsResponse', describeDBInstanceAutomatedBackupsResponse_httpStatus - The response's http status code.

Response Lenses

describeDBInstanceAutomatedBackupsResponse_dbInstanceAutomatedBackups :: Lens' DescribeDBInstanceAutomatedBackupsResponse (Maybe [DBInstanceAutomatedBackup])

A list of DBInstanceAutomatedBackup instances.

describeDBInstanceAutomatedBackupsResponse_marker :: Lens' DescribeDBInstanceAutomatedBackupsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

describeDBInstanceAutomatedBackupsResponse_httpStatus :: Lens' DescribeDBInstanceAutomatedBackupsResponse Int

The response's http status code.