Network.AWS.RDS.DescribeDBSnapshotAttributes

Description

Returns a list of DB snapshot attribute names and values for a manual DB snapshot.

When sharing snapshots with other Amazon Web Services accounts, DescribeDBSnapshotAttributes returns the restore attribute and a list of IDs for the Amazon Web Services accounts that are authorized to copy or restore the manual DB snapshot. If all is included in the list of values for the restore attribute, then the manual DB snapshot is public and can be copied or restored by all Amazon Web Services accounts.

To add or remove access for an Amazon Web Services account to copy or restore a manual DB snapshot, or to make the manual DB snapshot public or private, use the ModifyDBSnapshotAttribute API action.

Creating a Request

data DescribeDBSnapshotAttributes

See: newDescribeDBSnapshotAttributes smart constructor.

Constructors

DescribeDBSnapshotAttributes'

Fields

dbSnapshotIdentifier :: Text

The identifier for the DB snapshot to describe the attributes for.

newDescribeDBSnapshotAttributes

Arguments

:: Text
$sel:dbSnapshotIdentifier:DescribeDBSnapshotAttributes'

-> DescribeDBSnapshotAttributes

Create a value of DescribeDBSnapshotAttributes with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbSnapshotIdentifier:DescribeDBSnapshotAttributes', describeDBSnapshotAttributes_dbSnapshotIdentifier - The identifier for the DB snapshot to describe the attributes for.

Request Lenses

describeDBSnapshotAttributes_dbSnapshotIdentifier :: Lens' DescribeDBSnapshotAttributes Text

The identifier for the DB snapshot to describe the attributes for.

Destructuring the Response

data DescribeDBSnapshotAttributesResponse

See: newDescribeDBSnapshotAttributesResponse smart constructor.

Constructors

DescribeDBSnapshotAttributesResponse'

Fields

dbSnapshotAttributesResult :: Maybe DBSnapshotAttributesResult

httpStatus :: Int

The response's http status code.

newDescribeDBSnapshotAttributesResponse

Arguments

:: Int
$sel:httpStatus:DescribeDBSnapshotAttributesResponse'

-> DescribeDBSnapshotAttributesResponse

Create a value of DescribeDBSnapshotAttributesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbSnapshotAttributesResult:DescribeDBSnapshotAttributesResponse', describeDBSnapshotAttributesResponse_dbSnapshotAttributesResult - Undocumented member.

$sel:httpStatus:DescribeDBSnapshotAttributesResponse', describeDBSnapshotAttributesResponse_httpStatus - The response's http status code.

Response Lenses

describeDBSnapshotAttributesResponse_dbSnapshotAttributesResult :: Lens' DescribeDBSnapshotAttributesResponse (Maybe DBSnapshotAttributesResult)

Undocumented member.

describeDBSnapshotAttributesResponse_httpStatus :: Lens' DescribeDBSnapshotAttributesResponse Int

The response's http status code.