Network.AWS.RDS.DescribeDBSnapshots

Description

Returns information about DB snapshots. This API action supports pagination.

This operation returns paginated results.

Creating a Request

data DescribeDBSnapshots

See: newDescribeDBSnapshots smart constructor.

Constructors

DescribeDBSnapshots'

Fields

dbSnapshotIdentifier :: Maybe Text

A specific DB snapshot identifier to describe. This parameter can't be used in conjunction with DBInstanceIdentifier. This value is stored as a lowercase string.

Constraints:

If supplied, must match the identifier of an existing DBSnapshot.

If this identifier is for an automated snapshot, the SnapshotType parameter must also be specified.

dbiResourceId :: Maybe Text

A specific DB resource ID to describe.

includeShared :: Maybe Bool

A value that indicates whether to include shared manual DB cluster snapshots from other Amazon Web Services accounts that this Amazon Web Services account has been given permission to copy or restore. By default, these snapshots are not included.

You can give an Amazon Web Services account permission to restore a manual DB snapshot from another Amazon Web Services account by using the ModifyDBSnapshotAttribute API action.

snapshotType :: Maybe Text

The type of snapshots to be returned. You can specify one of the following values:

automated - Return all DB snapshots that have been automatically taken by Amazon RDS for my Amazon Web Services account.

manual - Return all DB snapshots that have been taken by my Amazon Web Services account.

shared - Return all manual DB snapshots that have been shared to my Amazon Web Services account.

public - Return all DB snapshots that have been marked as public.

awsbackup - Return the DB snapshots managed by the Amazon Web Services Backup service.

For information about Amazon Web Services Backup, see the Amazon Web Services Backup Developer Guide.

The awsbackup type does not apply to Aurora.

If you don't specify a SnapshotType value, then both automated and manual snapshots are returned. Shared and public DB snapshots are not included in the returned results by default. You can include shared snapshots with these results by enabling the IncludeShared parameter. You can include public snapshots with these results by enabling the IncludePublic parameter.

The IncludeShared and IncludePublic parameters don't apply for SnapshotType values of manual or automated. The IncludePublic parameter doesn't apply when SnapshotType is set to shared. The IncludeShared parameter doesn't apply when SnapshotType is set to public.

dbInstanceIdentifier :: Maybe Text

The ID of the DB instance to retrieve the list of DB snapshots for. This parameter can't be used in conjunction with DBSnapshotIdentifier. This parameter isn't case-sensitive.

Constraints:

If supplied, must match the identifier of an existing DBInstance.

filters :: Maybe [Filter]

A filter that specifies one or more DB snapshots to describe.

Supported filters:

db-instance-id - Accepts DB instance identifiers and DB instance Amazon Resource Names (ARNs).

db-snapshot-id - Accepts DB snapshot identifiers.

dbi-resource-id - Accepts identifiers of source DB instances.

snapshot-type - Accepts types of DB snapshots.

engine - Accepts names of database engines.

includePublic :: Maybe Bool

A value that indicates whether to include manual DB cluster snapshots that are public and can be copied or restored by any Amazon Web Services account. By default, the public snapshots are not included.

You can share a manual DB snapshot as public by using the ModifyDBSnapshotAttribute API.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

marker :: Maybe Text

An optional pagination token provided by a previous DescribeDBSnapshots request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

newDescribeDBSnapshots :: DescribeDBSnapshots

Create a value of DescribeDBSnapshots with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbSnapshotIdentifier:DescribeDBSnapshots', describeDBSnapshots_dbSnapshotIdentifier - A specific DB snapshot identifier to describe. This parameter can't be used in conjunction with DBInstanceIdentifier. This value is stored as a lowercase string.

Constraints:

If supplied, must match the identifier of an existing DBSnapshot.

If this identifier is for an automated snapshot, the SnapshotType parameter must also be specified.

$sel:dbiResourceId:DescribeDBSnapshots', describeDBSnapshots_dbiResourceId - A specific DB resource ID to describe.

$sel:includeShared:DescribeDBSnapshots', describeDBSnapshots_includeShared - A value that indicates whether to include shared manual DB cluster snapshots from other Amazon Web Services accounts that this Amazon Web Services account has been given permission to copy or restore. By default, these snapshots are not included.

You can give an Amazon Web Services account permission to restore a manual DB snapshot from another Amazon Web Services account by using the ModifyDBSnapshotAttribute API action.

$sel:snapshotType:DescribeDBSnapshots', describeDBSnapshots_snapshotType - The type of snapshots to be returned. You can specify one of the following values:

automated - Return all DB snapshots that have been automatically taken by Amazon RDS for my Amazon Web Services account.

manual - Return all DB snapshots that have been taken by my Amazon Web Services account.

shared - Return all manual DB snapshots that have been shared to my Amazon Web Services account.

public - Return all DB snapshots that have been marked as public.

awsbackup - Return the DB snapshots managed by the Amazon Web Services Backup service.

For information about Amazon Web Services Backup, see the Amazon Web Services Backup Developer Guide.

The awsbackup type does not apply to Aurora.

If you don't specify a SnapshotType value, then both automated and manual snapshots are returned. Shared and public DB snapshots are not included in the returned results by default. You can include shared snapshots with these results by enabling the IncludeShared parameter. You can include public snapshots with these results by enabling the IncludePublic parameter.

The IncludeShared and IncludePublic parameters don't apply for SnapshotType values of manual or automated. The IncludePublic parameter doesn't apply when SnapshotType is set to shared. The IncludeShared parameter doesn't apply when SnapshotType is set to public.

$sel:dbInstanceIdentifier:DescribeDBSnapshots', describeDBSnapshots_dbInstanceIdentifier - The ID of the DB instance to retrieve the list of DB snapshots for. This parameter can't be used in conjunction with DBSnapshotIdentifier. This parameter isn't case-sensitive.

Constraints:

If supplied, must match the identifier of an existing DBInstance.

$sel:filters:DescribeDBSnapshots', describeDBSnapshots_filters - A filter that specifies one or more DB snapshots to describe.

Supported filters:

db-instance-id - Accepts DB instance identifiers and DB instance Amazon Resource Names (ARNs).

db-snapshot-id - Accepts DB snapshot identifiers.

dbi-resource-id - Accepts identifiers of source DB instances.

snapshot-type - Accepts types of DB snapshots.

engine - Accepts names of database engines.

$sel:includePublic:DescribeDBSnapshots', describeDBSnapshots_includePublic - A value that indicates whether to include manual DB cluster snapshots that are public and can be copied or restored by any Amazon Web Services account. By default, the public snapshots are not included.

You can share a manual DB snapshot as public by using the ModifyDBSnapshotAttribute API.

$sel:maxRecords:DescribeDBSnapshots', describeDBSnapshots_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

$sel:marker:DescribeDBSnapshots', describeDBSnapshots_marker - An optional pagination token provided by a previous DescribeDBSnapshots request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Request Lenses

describeDBSnapshots_dbSnapshotIdentifier :: Lens' DescribeDBSnapshots (Maybe Text)

A specific DB snapshot identifier to describe. This parameter can't be used in conjunction with DBInstanceIdentifier. This value is stored as a lowercase string.

Constraints:

If supplied, must match the identifier of an existing DBSnapshot.

If this identifier is for an automated snapshot, the SnapshotType parameter must also be specified.

describeDBSnapshots_dbiResourceId :: Lens' DescribeDBSnapshots (Maybe Text)

A specific DB resource ID to describe.

describeDBSnapshots_includeShared :: Lens' DescribeDBSnapshots (Maybe Bool)

A value that indicates whether to include shared manual DB cluster snapshots from other Amazon Web Services accounts that this Amazon Web Services account has been given permission to copy or restore. By default, these snapshots are not included.

You can give an Amazon Web Services account permission to restore a manual DB snapshot from another Amazon Web Services account by using the ModifyDBSnapshotAttribute API action.

describeDBSnapshots_snapshotType :: Lens' DescribeDBSnapshots (Maybe Text)

The type of snapshots to be returned. You can specify one of the following values:

automated - Return all DB snapshots that have been automatically taken by Amazon RDS for my Amazon Web Services account.

manual - Return all DB snapshots that have been taken by my Amazon Web Services account.

shared - Return all manual DB snapshots that have been shared to my Amazon Web Services account.

public - Return all DB snapshots that have been marked as public.

awsbackup - Return the DB snapshots managed by the Amazon Web Services Backup service.

For information about Amazon Web Services Backup, see the Amazon Web Services Backup Developer Guide.

The awsbackup type does not apply to Aurora.

If you don't specify a SnapshotType value, then both automated and manual snapshots are returned. Shared and public DB snapshots are not included in the returned results by default. You can include shared snapshots with these results by enabling the IncludeShared parameter. You can include public snapshots with these results by enabling the IncludePublic parameter.

The IncludeShared and IncludePublic parameters don't apply for SnapshotType values of manual or automated. The IncludePublic parameter doesn't apply when SnapshotType is set to shared. The IncludeShared parameter doesn't apply when SnapshotType is set to public.

describeDBSnapshots_dbInstanceIdentifier :: Lens' DescribeDBSnapshots (Maybe Text)

The ID of the DB instance to retrieve the list of DB snapshots for. This parameter can't be used in conjunction with DBSnapshotIdentifier. This parameter isn't case-sensitive.

Constraints:

If supplied, must match the identifier of an existing DBInstance.

describeDBSnapshots_filters :: Lens' DescribeDBSnapshots (Maybe [Filter])

A filter that specifies one or more DB snapshots to describe.

Supported filters:

db-instance-id - Accepts DB instance identifiers and DB instance Amazon Resource Names (ARNs).

db-snapshot-id - Accepts DB snapshot identifiers.

dbi-resource-id - Accepts identifiers of source DB instances.

snapshot-type - Accepts types of DB snapshots.

engine - Accepts names of database engines.

describeDBSnapshots_includePublic :: Lens' DescribeDBSnapshots (Maybe Bool)

A value that indicates whether to include manual DB cluster snapshots that are public and can be copied or restored by any Amazon Web Services account. By default, the public snapshots are not included.

You can share a manual DB snapshot as public by using the ModifyDBSnapshotAttribute API.

describeDBSnapshots_maxRecords :: Lens' DescribeDBSnapshots (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

describeDBSnapshots_marker :: Lens' DescribeDBSnapshots (Maybe Text)

An optional pagination token provided by a previous DescribeDBSnapshots request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Destructuring the Response

data DescribeDBSnapshotsResponse

Contains the result of a successful invocation of the DescribeDBSnapshots action.

See: newDescribeDBSnapshotsResponse smart constructor.

Constructors

DescribeDBSnapshotsResponse'

Fields

dbSnapshots :: Maybe [DBSnapshot]

A list of DBSnapshot instances.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

httpStatus :: Int

The response's http status code.

newDescribeDBSnapshotsResponse

Arguments

:: Int
$sel:httpStatus:DescribeDBSnapshotsResponse'

-> DescribeDBSnapshotsResponse

Create a value of DescribeDBSnapshotsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbSnapshots:DescribeDBSnapshotsResponse', describeDBSnapshotsResponse_dbSnapshots - A list of DBSnapshot instances.

$sel:marker:DescribeDBSnapshots', describeDBSnapshotsResponse_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:httpStatus:DescribeDBSnapshotsResponse', describeDBSnapshotsResponse_httpStatus - The response's http status code.

Response Lenses

describeDBSnapshotsResponse_dbSnapshots :: Lens' DescribeDBSnapshotsResponse (Maybe [DBSnapshot])

A list of DBSnapshot instances.

describeDBSnapshotsResponse_marker :: Lens' DescribeDBSnapshotsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeDBSnapshotsResponse_httpStatus :: Lens' DescribeDBSnapshotsResponse Int

The response's http status code.