Network.AWS.RDS.DescribeGlobalClusters

Description

Returns information about Aurora global database clusters. This API supports pagination.

For more information on Amazon Aurora, see What Is Amazon Aurora? in the Amazon Aurora User Guide.

This action only applies to Aurora DB clusters.

This operation returns paginated results.

Creating a Request

data DescribeGlobalClusters

See: newDescribeGlobalClusters smart constructor.

Constructors

DescribeGlobalClusters'

Fields

filters :: Maybe [Filter]

This parameter isn't currently supported.

globalClusterIdentifier :: Maybe Text

The user-supplied DB cluster identifier. If this parameter is specified, information from only the specific DB cluster is returned. This parameter isn't case-sensitive.

Constraints:

If supplied, must match an existing DBClusterIdentifier.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

marker :: Maybe Text

An optional pagination token provided by a previous DescribeGlobalClusters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

newDescribeGlobalClusters :: DescribeGlobalClusters

Create a value of DescribeGlobalClusters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeGlobalClusters', describeGlobalClusters_filters - This parameter isn't currently supported.

$sel:globalClusterIdentifier:DescribeGlobalClusters', describeGlobalClusters_globalClusterIdentifier - The user-supplied DB cluster identifier. If this parameter is specified, information from only the specific DB cluster is returned. This parameter isn't case-sensitive.

Constraints:

If supplied, must match an existing DBClusterIdentifier.

$sel:maxRecords:DescribeGlobalClusters', describeGlobalClusters_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

$sel:marker:DescribeGlobalClusters', describeGlobalClusters_marker - An optional pagination token provided by a previous DescribeGlobalClusters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Request Lenses

describeGlobalClusters_filters :: Lens' DescribeGlobalClusters (Maybe [Filter])

This parameter isn't currently supported.

describeGlobalClusters_globalClusterIdentifier :: Lens' DescribeGlobalClusters (Maybe Text)

The user-supplied DB cluster identifier. If this parameter is specified, information from only the specific DB cluster is returned. This parameter isn't case-sensitive.

Constraints:

If supplied, must match an existing DBClusterIdentifier.

describeGlobalClusters_maxRecords :: Lens' DescribeGlobalClusters (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

describeGlobalClusters_marker :: Lens' DescribeGlobalClusters (Maybe Text)

An optional pagination token provided by a previous DescribeGlobalClusters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Destructuring the Response

data DescribeGlobalClustersResponse

See: newDescribeGlobalClustersResponse smart constructor.

Constructors

DescribeGlobalClustersResponse'

Fields

globalClusters :: Maybe [GlobalCluster]

The list of global clusters returned by this request.

marker :: Maybe Text

An optional pagination token provided by a previous DescribeGlobalClusters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

httpStatus :: Int

The response's http status code.

newDescribeGlobalClustersResponse

Arguments

:: Int
$sel:httpStatus:DescribeGlobalClustersResponse'

-> DescribeGlobalClustersResponse

Create a value of DescribeGlobalClustersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:globalClusters:DescribeGlobalClustersResponse', describeGlobalClustersResponse_globalClusters - The list of global clusters returned by this request.

$sel:marker:DescribeGlobalClusters', describeGlobalClustersResponse_marker - An optional pagination token provided by a previous DescribeGlobalClusters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:httpStatus:DescribeGlobalClustersResponse', describeGlobalClustersResponse_httpStatus - The response's http status code.

Response Lenses

describeGlobalClustersResponse_globalClusters :: Lens' DescribeGlobalClustersResponse (Maybe [GlobalCluster])

The list of global clusters returned by this request.

describeGlobalClustersResponse_marker :: Lens' DescribeGlobalClustersResponse (Maybe Text)

An optional pagination token provided by a previous DescribeGlobalClusters request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeGlobalClustersResponse_httpStatus :: Lens' DescribeGlobalClustersResponse Int

The response's http status code.