Network.AWS.RDS.DescribeInstallationMedia

Description

Describes the available installation media for a DB engine that requires an on-premises customer provided license, such as Microsoft SQL Server.

This operation returns paginated results.

Creating a Request

data DescribeInstallationMedia

See: newDescribeInstallationMedia smart constructor.

Constructors

DescribeInstallationMedia'

Fields

installationMediaId :: Maybe Text

The installation medium ID.

filters :: Maybe [Filter]

A filter that specifies one or more installation media to describe. Supported filters include the following:

custom-availability-zone-id - Accepts custom Availability Zone (AZ) identifiers. The results list includes information about only the custom AZs identified by these identifiers.

engine - Accepts database engines. The results list includes information about only the database engines identified by these identifiers.

For more information about the valid engines for installation media, see ImportInstallationMedia.

maxRecords :: Maybe Int

An optional pagination token provided by a previous DescribeInstallationMedia request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

newDescribeInstallationMedia :: DescribeInstallationMedia

Create a value of DescribeInstallationMedia with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:installationMediaId:DescribeInstallationMedia', describeInstallationMedia_installationMediaId - The installation medium ID.

$sel:filters:DescribeInstallationMedia', describeInstallationMedia_filters - A filter that specifies one or more installation media to describe. Supported filters include the following:

custom-availability-zone-id - Accepts custom Availability Zone (AZ) identifiers. The results list includes information about only the custom AZs identified by these identifiers.

engine - Accepts database engines. The results list includes information about only the database engines identified by these identifiers.

For more information about the valid engines for installation media, see ImportInstallationMedia.

$sel:maxRecords:DescribeInstallationMedia', describeInstallationMedia_maxRecords - An optional pagination token provided by a previous DescribeInstallationMedia request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:marker:DescribeInstallationMedia', describeInstallationMedia_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Request Lenses

describeInstallationMedia_installationMediaId :: Lens' DescribeInstallationMedia (Maybe Text)

The installation medium ID.

describeInstallationMedia_filters :: Lens' DescribeInstallationMedia (Maybe [Filter])

A filter that specifies one or more installation media to describe. Supported filters include the following:

custom-availability-zone-id - Accepts custom Availability Zone (AZ) identifiers. The results list includes information about only the custom AZs identified by these identifiers.

engine - Accepts database engines. The results list includes information about only the database engines identified by these identifiers.

For more information about the valid engines for installation media, see ImportInstallationMedia.

describeInstallationMedia_maxRecords :: Lens' DescribeInstallationMedia (Maybe Int)

An optional pagination token provided by a previous DescribeInstallationMedia request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeInstallationMedia_marker :: Lens' DescribeInstallationMedia (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Destructuring the Response

data DescribeInstallationMediaResponse

See: newDescribeInstallationMediaResponse smart constructor.

Constructors

DescribeInstallationMediaResponse'

Fields

installationMedia :: Maybe [InstallationMedia]

The list of InstallationMedia objects for the Amazon Web Services account.

marker :: Maybe Text

An optional pagination token provided by a previous DescribeInstallationMedia request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

httpStatus :: Int

The response's http status code.

newDescribeInstallationMediaResponse

Arguments

:: Int
$sel:httpStatus:DescribeInstallationMediaResponse'

-> DescribeInstallationMediaResponse

Create a value of DescribeInstallationMediaResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:installationMedia:DescribeInstallationMediaResponse', describeInstallationMediaResponse_installationMedia - The list of InstallationMedia objects for the Amazon Web Services account.

$sel:marker:DescribeInstallationMedia', describeInstallationMediaResponse_marker - An optional pagination token provided by a previous DescribeInstallationMedia request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:httpStatus:DescribeInstallationMediaResponse', describeInstallationMediaResponse_httpStatus - The response's http status code.

Response Lenses

describeInstallationMediaResponse_installationMedia :: Lens' DescribeInstallationMediaResponse (Maybe [InstallationMedia])

The list of InstallationMedia objects for the Amazon Web Services account.

describeInstallationMediaResponse_marker :: Lens' DescribeInstallationMediaResponse (Maybe Text)

An optional pagination token provided by a previous DescribeInstallationMedia request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeInstallationMediaResponse_httpStatus :: Lens' DescribeInstallationMediaResponse Int

The response's http status code.