Network.AWS.RDS.DescribeOptionGroupOptions

Description

Describes all available options.

This operation returns paginated results.

Creating a Request

data DescribeOptionGroupOptions

See: newDescribeOptionGroupOptions smart constructor.

Constructors

DescribeOptionGroupOptions'

Fields

filters :: Maybe [Filter]

This parameter isn't currently supported.

majorEngineVersion :: Maybe Text

If specified, filters the results to include only options for the specified major engine version.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

engineName :: Text

A required parameter. Options available for the given engine name are described.

Valid Values:

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

newDescribeOptionGroupOptions

Arguments

:: Text
$sel:engineName:DescribeOptionGroupOptions'

-> DescribeOptionGroupOptions

Create a value of DescribeOptionGroupOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:filters:DescribeOptionGroupOptions', describeOptionGroupOptions_filters - This parameter isn't currently supported.

$sel:majorEngineVersion:DescribeOptionGroupOptions', describeOptionGroupOptions_majorEngineVersion - If specified, filters the results to include only options for the specified major engine version.

$sel:marker:DescribeOptionGroupOptions', describeOptionGroupOptions_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:maxRecords:DescribeOptionGroupOptions', describeOptionGroupOptions_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

$sel:engineName:DescribeOptionGroupOptions', describeOptionGroupOptions_engineName - A required parameter. Options available for the given engine name are described.

Valid Values:

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

Request Lenses

describeOptionGroupOptions_filters :: Lens' DescribeOptionGroupOptions (Maybe [Filter])

This parameter isn't currently supported.

describeOptionGroupOptions_majorEngineVersion :: Lens' DescribeOptionGroupOptions (Maybe Text)

If specified, filters the results to include only options for the specified major engine version.

describeOptionGroupOptions_marker :: Lens' DescribeOptionGroupOptions (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeOptionGroupOptions_maxRecords :: Lens' DescribeOptionGroupOptions (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

describeOptionGroupOptions_engineName :: Lens' DescribeOptionGroupOptions Text

A required parameter. Options available for the given engine name are described.

Valid Values:

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

Destructuring the Response

data DescribeOptionGroupOptionsResponse

See: newDescribeOptionGroupOptionsResponse smart constructor.

Constructors

DescribeOptionGroupOptionsResponse'

Fields

optionGroupOptions :: Maybe [OptionGroupOption]

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

httpStatus :: Int

The response's http status code.

newDescribeOptionGroupOptionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeOptionGroupOptionsResponse'

-> DescribeOptionGroupOptionsResponse

Create a value of DescribeOptionGroupOptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:optionGroupOptions:DescribeOptionGroupOptionsResponse', describeOptionGroupOptionsResponse_optionGroupOptions - Undocumented member.

$sel:marker:DescribeOptionGroupOptions', describeOptionGroupOptionsResponse_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:httpStatus:DescribeOptionGroupOptionsResponse', describeOptionGroupOptionsResponse_httpStatus - The response's http status code.

Response Lenses

describeOptionGroupOptionsResponse_optionGroupOptions :: Lens' DescribeOptionGroupOptionsResponse (Maybe [OptionGroupOption])

Undocumented member.

describeOptionGroupOptionsResponse_marker :: Lens' DescribeOptionGroupOptionsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeOptionGroupOptionsResponse_httpStatus :: Lens' DescribeOptionGroupOptionsResponse Int

The response's http status code.