Network.AWS.RDS.DescribeOrderableDBInstanceOptions

Description

Returns a list of orderable DB instance options for the specified engine.

This operation returns paginated results.

Creating a Request

data DescribeOrderableDBInstanceOptions

See: newDescribeOrderableDBInstanceOptions smart constructor.

Constructors

DescribeOrderableDBInstanceOptions'

Fields

engineVersion :: Maybe Text

The engine version filter value. Specify this parameter to show only the available offerings matching the specified engine version.

licenseModel :: Maybe Text

The license model filter value. Specify this parameter to show only the available offerings matching the specified license model.

dbInstanceClass :: Maybe Text

The DB instance class filter value. Specify this parameter to show only the available offerings matching the specified DB instance class.

filters :: Maybe [Filter]

This parameter isn't currently supported.

availabilityZoneGroup :: Maybe Text

The Availability Zone group associated with a Local Zone. Specify this parameter to retrieve available offerings for the Local Zones in the group.

Omit this parameter to show the available offerings in the specified Amazon Web Services Region.

vpc :: Maybe Bool

A value that indicates whether to show only VPC or non-VPC offerings.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

marker :: Maybe Text

An optional pagination token provided by a previous DescribeOrderableDBInstanceOptions request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

engine :: Text

The name of the engine to retrieve DB instance options for.

Valid Values:

aurora (for MySQL 5.6-compatible Aurora)

aurora-mysql (for MySQL 5.7-compatible Aurora)

aurora-postgresql

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

newDescribeOrderableDBInstanceOptions

Arguments

:: Text
$sel:engine:DescribeOrderableDBInstanceOptions'

-> DescribeOrderableDBInstanceOptions

Create a value of DescribeOrderableDBInstanceOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:DescribeOrderableDBInstanceOptions', describeOrderableDBInstanceOptions_engineVersion - The engine version filter value. Specify this parameter to show only the available offerings matching the specified engine version.

$sel:licenseModel:DescribeOrderableDBInstanceOptions', describeOrderableDBInstanceOptions_licenseModel - The license model filter value. Specify this parameter to show only the available offerings matching the specified license model.

$sel:dbInstanceClass:DescribeOrderableDBInstanceOptions', describeOrderableDBInstanceOptions_dbInstanceClass - The DB instance class filter value. Specify this parameter to show only the available offerings matching the specified DB instance class.

$sel:filters:DescribeOrderableDBInstanceOptions', describeOrderableDBInstanceOptions_filters - This parameter isn't currently supported.

$sel:availabilityZoneGroup:DescribeOrderableDBInstanceOptions', describeOrderableDBInstanceOptions_availabilityZoneGroup - The Availability Zone group associated with a Local Zone. Specify this parameter to retrieve available offerings for the Local Zones in the group.

Omit this parameter to show the available offerings in the specified Amazon Web Services Region.

$sel:vpc:DescribeOrderableDBInstanceOptions', describeOrderableDBInstanceOptions_vpc - A value that indicates whether to show only VPC or non-VPC offerings.

$sel:maxRecords:DescribeOrderableDBInstanceOptions', describeOrderableDBInstanceOptions_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

$sel:marker:DescribeOrderableDBInstanceOptions', describeOrderableDBInstanceOptions_marker - An optional pagination token provided by a previous DescribeOrderableDBInstanceOptions request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

$sel:engine:DescribeOrderableDBInstanceOptions', describeOrderableDBInstanceOptions_engine - The name of the engine to retrieve DB instance options for.

Valid Values:

aurora (for MySQL 5.6-compatible Aurora)

aurora-mysql (for MySQL 5.7-compatible Aurora)

aurora-postgresql

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

Request Lenses

describeOrderableDBInstanceOptions_engineVersion :: Lens' DescribeOrderableDBInstanceOptions (Maybe Text)

The engine version filter value. Specify this parameter to show only the available offerings matching the specified engine version.

describeOrderableDBInstanceOptions_licenseModel :: Lens' DescribeOrderableDBInstanceOptions (Maybe Text)

The license model filter value. Specify this parameter to show only the available offerings matching the specified license model.

describeOrderableDBInstanceOptions_dbInstanceClass :: Lens' DescribeOrderableDBInstanceOptions (Maybe Text)

The DB instance class filter value. Specify this parameter to show only the available offerings matching the specified DB instance class.

describeOrderableDBInstanceOptions_filters :: Lens' DescribeOrderableDBInstanceOptions (Maybe [Filter])

This parameter isn't currently supported.

describeOrderableDBInstanceOptions_availabilityZoneGroup :: Lens' DescribeOrderableDBInstanceOptions (Maybe Text)

The Availability Zone group associated with a Local Zone. Specify this parameter to retrieve available offerings for the Local Zones in the group.

Omit this parameter to show the available offerings in the specified Amazon Web Services Region.

describeOrderableDBInstanceOptions_vpc :: Lens' DescribeOrderableDBInstanceOptions (Maybe Bool)

A value that indicates whether to show only VPC or non-VPC offerings.

describeOrderableDBInstanceOptions_maxRecords :: Lens' DescribeOrderableDBInstanceOptions (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

describeOrderableDBInstanceOptions_marker :: Lens' DescribeOrderableDBInstanceOptions (Maybe Text)

An optional pagination token provided by a previous DescribeOrderableDBInstanceOptions request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

describeOrderableDBInstanceOptions_engine :: Lens' DescribeOrderableDBInstanceOptions Text

The name of the engine to retrieve DB instance options for.

Valid Values:

aurora (for MySQL 5.6-compatible Aurora)

aurora-mysql (for MySQL 5.7-compatible Aurora)

aurora-postgresql

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

Destructuring the Response

data DescribeOrderableDBInstanceOptionsResponse

Contains the result of a successful invocation of the DescribeOrderableDBInstanceOptions action.

See: newDescribeOrderableDBInstanceOptionsResponse smart constructor.

Constructors

DescribeOrderableDBInstanceOptionsResponse'

Fields

orderableDBInstanceOptions :: Maybe [OrderableDBInstanceOption]

An OrderableDBInstanceOption structure containing information about orderable options for the DB instance.

marker :: Maybe Text

An optional pagination token provided by a previous OrderableDBInstanceOptions request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

httpStatus :: Int

The response's http status code.

newDescribeOrderableDBInstanceOptionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeOrderableDBInstanceOptionsResponse'

-> DescribeOrderableDBInstanceOptionsResponse

Create a value of DescribeOrderableDBInstanceOptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:orderableDBInstanceOptions:DescribeOrderableDBInstanceOptionsResponse', describeOrderableDBInstanceOptionsResponse_orderableDBInstanceOptions - An OrderableDBInstanceOption structure containing information about orderable options for the DB instance.

$sel:marker:DescribeOrderableDBInstanceOptions', describeOrderableDBInstanceOptionsResponse_marker - An optional pagination token provided by a previous OrderableDBInstanceOptions request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

$sel:httpStatus:DescribeOrderableDBInstanceOptionsResponse', describeOrderableDBInstanceOptionsResponse_httpStatus - The response's http status code.

Response Lenses

describeOrderableDBInstanceOptionsResponse_orderableDBInstanceOptions :: Lens' DescribeOrderableDBInstanceOptionsResponse (Maybe [OrderableDBInstanceOption])

An OrderableDBInstanceOption structure containing information about orderable options for the DB instance.

describeOrderableDBInstanceOptionsResponse_marker :: Lens' DescribeOrderableDBInstanceOptionsResponse (Maybe Text)

An optional pagination token provided by a previous OrderableDBInstanceOptions request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

describeOrderableDBInstanceOptionsResponse_httpStatus :: Lens' DescribeOrderableDBInstanceOptionsResponse Int

The response's http status code.