Network.AWS.RDS.DescribeReservedDBInstances

Description

Returns information about reserved DB instances for this account, or about a specified reserved DB instance.

This operation returns paginated results.

Creating a Request

data DescribeReservedDBInstances

See: newDescribeReservedDBInstances smart constructor.

Constructors

DescribeReservedDBInstances'

Fields

duration :: Maybe Text

The duration filter value, specified in years or seconds. Specify this parameter to show only reservations for this duration.

Valid Values: 1 | 3 | 31536000 | 94608000

multiAZ :: Maybe Bool

A value that indicates whether to show only those reservations that support Multi-AZ.

dbInstanceClass :: Maybe Text

The DB instance class filter value. Specify this parameter to show only those reservations matching the specified DB instances class.

reservedDBInstanceId :: Maybe Text

The reserved DB instance identifier filter value. Specify this parameter to show only the reservation that matches the specified reservation ID.

leaseId :: Maybe Text

The lease identifier filter value. Specify this parameter to show only the reservation that matches the specified lease ID.

Amazon Web Services Support might request the lease ID for an issue related to a reserved DB instance.

filters :: Maybe [Filter]

This parameter isn't currently supported.

offeringType :: Maybe Text

The offering type filter value. Specify this parameter to show only the available offerings matching the specified offering type.

Valid Values: "Partial Upfront" | "All Upfront" | "No Upfront"

productDescription :: Maybe Text

The product description filter value. Specify this parameter to show only those reservations matching the specified product description.

reservedDBInstancesOfferingId :: Maybe Text

The offering identifier filter value. Specify this parameter to show only purchased reservations matching the specified offering identifier.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more than the MaxRecords value is available, a pagination token called a marker is included in the response so you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

newDescribeReservedDBInstances :: DescribeReservedDBInstances

Create a value of DescribeReservedDBInstances with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:duration:DescribeReservedDBInstances', describeReservedDBInstances_duration - The duration filter value, specified in years or seconds. Specify this parameter to show only reservations for this duration.

Valid Values: 1 | 3 | 31536000 | 94608000

$sel:multiAZ:DescribeReservedDBInstances', describeReservedDBInstances_multiAZ - A value that indicates whether to show only those reservations that support Multi-AZ.

$sel:dbInstanceClass:DescribeReservedDBInstances', describeReservedDBInstances_dbInstanceClass - The DB instance class filter value. Specify this parameter to show only those reservations matching the specified DB instances class.

$sel:reservedDBInstanceId:DescribeReservedDBInstances', describeReservedDBInstances_reservedDBInstanceId - The reserved DB instance identifier filter value. Specify this parameter to show only the reservation that matches the specified reservation ID.

$sel:leaseId:DescribeReservedDBInstances', describeReservedDBInstances_leaseId - The lease identifier filter value. Specify this parameter to show only the reservation that matches the specified lease ID.

Amazon Web Services Support might request the lease ID for an issue related to a reserved DB instance.

$sel:filters:DescribeReservedDBInstances', describeReservedDBInstances_filters - This parameter isn't currently supported.

$sel:offeringType:DescribeReservedDBInstances', describeReservedDBInstances_offeringType - The offering type filter value. Specify this parameter to show only the available offerings matching the specified offering type.

Valid Values: "Partial Upfront" | "All Upfront" | "No Upfront"

$sel:productDescription:DescribeReservedDBInstances', describeReservedDBInstances_productDescription - The product description filter value. Specify this parameter to show only those reservations matching the specified product description.

$sel:reservedDBInstancesOfferingId:DescribeReservedDBInstances', describeReservedDBInstances_reservedDBInstancesOfferingId - The offering identifier filter value. Specify this parameter to show only purchased reservations matching the specified offering identifier.

$sel:maxRecords:DescribeReservedDBInstances', describeReservedDBInstances_maxRecords - The maximum number of records to include in the response. If more than the MaxRecords value is available, a pagination token called a marker is included in the response so you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

$sel:marker:DescribeReservedDBInstances', describeReservedDBInstances_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Request Lenses

describeReservedDBInstances_duration :: Lens' DescribeReservedDBInstances (Maybe Text)

The duration filter value, specified in years or seconds. Specify this parameter to show only reservations for this duration.

Valid Values: 1 | 3 | 31536000 | 94608000

describeReservedDBInstances_multiAZ :: Lens' DescribeReservedDBInstances (Maybe Bool)

A value that indicates whether to show only those reservations that support Multi-AZ.

describeReservedDBInstances_dbInstanceClass :: Lens' DescribeReservedDBInstances (Maybe Text)

The DB instance class filter value. Specify this parameter to show only those reservations matching the specified DB instances class.

describeReservedDBInstances_reservedDBInstanceId :: Lens' DescribeReservedDBInstances (Maybe Text)

The reserved DB instance identifier filter value. Specify this parameter to show only the reservation that matches the specified reservation ID.

describeReservedDBInstances_leaseId :: Lens' DescribeReservedDBInstances (Maybe Text)

The lease identifier filter value. Specify this parameter to show only the reservation that matches the specified lease ID.

Amazon Web Services Support might request the lease ID for an issue related to a reserved DB instance.

describeReservedDBInstances_filters :: Lens' DescribeReservedDBInstances (Maybe [Filter])

This parameter isn't currently supported.

describeReservedDBInstances_offeringType :: Lens' DescribeReservedDBInstances (Maybe Text)

The offering type filter value. Specify this parameter to show only the available offerings matching the specified offering type.

Valid Values: "Partial Upfront" | "All Upfront" | "No Upfront"

describeReservedDBInstances_productDescription :: Lens' DescribeReservedDBInstances (Maybe Text)

The product description filter value. Specify this parameter to show only those reservations matching the specified product description.

describeReservedDBInstances_reservedDBInstancesOfferingId :: Lens' DescribeReservedDBInstances (Maybe Text)

The offering identifier filter value. Specify this parameter to show only purchased reservations matching the specified offering identifier.

describeReservedDBInstances_maxRecords :: Lens' DescribeReservedDBInstances (Maybe Int)

The maximum number of records to include in the response. If more than the MaxRecords value is available, a pagination token called a marker is included in the response so you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

describeReservedDBInstances_marker :: Lens' DescribeReservedDBInstances (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Destructuring the Response

data DescribeReservedDBInstancesResponse

Contains the result of a successful invocation of the DescribeReservedDBInstances action.

See: newDescribeReservedDBInstancesResponse smart constructor.

Constructors

DescribeReservedDBInstancesResponse'

Fields

reservedDBInstances :: Maybe [ReservedDBInstance]

A list of reserved DB instances.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

httpStatus :: Int

The response's http status code.

newDescribeReservedDBInstancesResponse

Arguments

:: Int
$sel:httpStatus:DescribeReservedDBInstancesResponse'

-> DescribeReservedDBInstancesResponse

Create a value of DescribeReservedDBInstancesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reservedDBInstances:DescribeReservedDBInstancesResponse', describeReservedDBInstancesResponse_reservedDBInstances - A list of reserved DB instances.

$sel:marker:DescribeReservedDBInstances', describeReservedDBInstancesResponse_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:httpStatus:DescribeReservedDBInstancesResponse', describeReservedDBInstancesResponse_httpStatus - The response's http status code.

Response Lenses

describeReservedDBInstancesResponse_reservedDBInstances :: Lens' DescribeReservedDBInstancesResponse (Maybe [ReservedDBInstance])

A list of reserved DB instances.

describeReservedDBInstancesResponse_marker :: Lens' DescribeReservedDBInstancesResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeReservedDBInstancesResponse_httpStatus :: Lens' DescribeReservedDBInstancesResponse Int

The response's http status code.