Network.AWS.RDS.DescribeReservedDBInstancesOfferings

Description

Lists available reserved DB instance offerings.

This operation returns paginated results.

Creating a Request

data DescribeReservedDBInstancesOfferings

See: newDescribeReservedDBInstancesOfferings smart constructor.

Constructors

DescribeReservedDBInstancesOfferings'

Fields

duration :: Maybe Text

Duration filter value, specified in years or seconds. Specify this parameter to show only reservations for this duration.

Valid Values: 1 | 3 | 31536000 | 94608000

multiAZ :: Maybe Bool

A value that indicates whether to show only those reservations that support Multi-AZ.

dbInstanceClass :: Maybe Text

The DB instance class filter value. Specify this parameter to show only the available offerings matching the specified DB instance class.

filters :: Maybe [Filter]

This parameter isn't currently supported.

offeringType :: Maybe Text

The offering type filter value. Specify this parameter to show only the available offerings matching the specified offering type.

Valid Values: "Partial Upfront" | "All Upfront" | "No Upfront"

productDescription :: Maybe Text

Product description filter value. Specify this parameter to show only the available offerings that contain the specified product description.

The results show offerings that partially match the filter value.

reservedDBInstancesOfferingId :: Maybe Text

The offering identifier filter value. Specify this parameter to show only the available offering that matches the specified reservation identifier.

Example: 438012d3-4052-4cc7-b2e3-8d3372e0e706

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more than the MaxRecords value is available, a pagination token called a marker is included in the response so you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

newDescribeReservedDBInstancesOfferings :: DescribeReservedDBInstancesOfferings

Create a value of DescribeReservedDBInstancesOfferings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:duration:DescribeReservedDBInstancesOfferings', describeReservedDBInstancesOfferings_duration - Duration filter value, specified in years or seconds. Specify this parameter to show only reservations for this duration.

Valid Values: 1 | 3 | 31536000 | 94608000

$sel:multiAZ:DescribeReservedDBInstancesOfferings', describeReservedDBInstancesOfferings_multiAZ - A value that indicates whether to show only those reservations that support Multi-AZ.

$sel:dbInstanceClass:DescribeReservedDBInstancesOfferings', describeReservedDBInstancesOfferings_dbInstanceClass - The DB instance class filter value. Specify this parameter to show only the available offerings matching the specified DB instance class.

$sel:filters:DescribeReservedDBInstancesOfferings', describeReservedDBInstancesOfferings_filters - This parameter isn't currently supported.

$sel:offeringType:DescribeReservedDBInstancesOfferings', describeReservedDBInstancesOfferings_offeringType - The offering type filter value. Specify this parameter to show only the available offerings matching the specified offering type.

Valid Values: "Partial Upfront" | "All Upfront" | "No Upfront"

$sel:productDescription:DescribeReservedDBInstancesOfferings', describeReservedDBInstancesOfferings_productDescription - Product description filter value. Specify this parameter to show only the available offerings that contain the specified product description.

The results show offerings that partially match the filter value.

$sel:reservedDBInstancesOfferingId:DescribeReservedDBInstancesOfferings', describeReservedDBInstancesOfferings_reservedDBInstancesOfferingId - The offering identifier filter value. Specify this parameter to show only the available offering that matches the specified reservation identifier.

Example: 438012d3-4052-4cc7-b2e3-8d3372e0e706

$sel:maxRecords:DescribeReservedDBInstancesOfferings', describeReservedDBInstancesOfferings_maxRecords - The maximum number of records to include in the response. If more than the MaxRecords value is available, a pagination token called a marker is included in the response so you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

$sel:marker:DescribeReservedDBInstancesOfferings', describeReservedDBInstancesOfferings_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Request Lenses

describeReservedDBInstancesOfferings_duration :: Lens' DescribeReservedDBInstancesOfferings (Maybe Text)

Duration filter value, specified in years or seconds. Specify this parameter to show only reservations for this duration.

Valid Values: 1 | 3 | 31536000 | 94608000

describeReservedDBInstancesOfferings_multiAZ :: Lens' DescribeReservedDBInstancesOfferings (Maybe Bool)

A value that indicates whether to show only those reservations that support Multi-AZ.

describeReservedDBInstancesOfferings_dbInstanceClass :: Lens' DescribeReservedDBInstancesOfferings (Maybe Text)

The DB instance class filter value. Specify this parameter to show only the available offerings matching the specified DB instance class.

describeReservedDBInstancesOfferings_filters :: Lens' DescribeReservedDBInstancesOfferings (Maybe [Filter])

This parameter isn't currently supported.

describeReservedDBInstancesOfferings_offeringType :: Lens' DescribeReservedDBInstancesOfferings (Maybe Text)

The offering type filter value. Specify this parameter to show only the available offerings matching the specified offering type.

Valid Values: "Partial Upfront" | "All Upfront" | "No Upfront"

describeReservedDBInstancesOfferings_productDescription :: Lens' DescribeReservedDBInstancesOfferings (Maybe Text)

Product description filter value. Specify this parameter to show only the available offerings that contain the specified product description.

The results show offerings that partially match the filter value.

describeReservedDBInstancesOfferings_reservedDBInstancesOfferingId :: Lens' DescribeReservedDBInstancesOfferings (Maybe Text)

The offering identifier filter value. Specify this parameter to show only the available offering that matches the specified reservation identifier.

Example: 438012d3-4052-4cc7-b2e3-8d3372e0e706

describeReservedDBInstancesOfferings_maxRecords :: Lens' DescribeReservedDBInstancesOfferings (Maybe Int)

The maximum number of records to include in the response. If more than the MaxRecords value is available, a pagination token called a marker is included in the response so you can retrieve the remaining results.

Default: 100

Constraints: Minimum 20, maximum 100.

describeReservedDBInstancesOfferings_marker :: Lens' DescribeReservedDBInstancesOfferings (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

Destructuring the Response

data DescribeReservedDBInstancesOfferingsResponse

Contains the result of a successful invocation of the DescribeReservedDBInstancesOfferings action.

See: newDescribeReservedDBInstancesOfferingsResponse smart constructor.

Constructors

DescribeReservedDBInstancesOfferingsResponse'

Fields

reservedDBInstancesOfferings :: Maybe [ReservedDBInstancesOffering]

A list of reserved DB instance offerings.

marker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

httpStatus :: Int

The response's http status code.

newDescribeReservedDBInstancesOfferingsResponse

Arguments

:: Int
$sel:httpStatus:DescribeReservedDBInstancesOfferingsResponse'

-> DescribeReservedDBInstancesOfferingsResponse

Create a value of DescribeReservedDBInstancesOfferingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reservedDBInstancesOfferings:DescribeReservedDBInstancesOfferingsResponse', describeReservedDBInstancesOfferingsResponse_reservedDBInstancesOfferings - A list of reserved DB instance offerings.

$sel:marker:DescribeReservedDBInstancesOfferings', describeReservedDBInstancesOfferingsResponse_marker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

$sel:httpStatus:DescribeReservedDBInstancesOfferingsResponse', describeReservedDBInstancesOfferingsResponse_httpStatus - The response's http status code.

Response Lenses

describeReservedDBInstancesOfferingsResponse_reservedDBInstancesOfferings :: Lens' DescribeReservedDBInstancesOfferingsResponse (Maybe [ReservedDBInstancesOffering])

A list of reserved DB instance offerings.

describeReservedDBInstancesOfferingsResponse_marker :: Lens' DescribeReservedDBInstancesOfferingsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

describeReservedDBInstancesOfferingsResponse_httpStatus :: Lens' DescribeReservedDBInstancesOfferingsResponse Int

The response's http status code.