Network.AWS.RDS.ImportInstallationMedia

Description

Imports the installation media for a DB engine that requires an on-premises customer provided license, such as SQL Server.

Creating a Request

data ImportInstallationMedia

See: newImportInstallationMedia smart constructor.

Constructors

ImportInstallationMedia'

Fields

customAvailabilityZoneId :: Text

The identifier of the custom Availability Zone (AZ) to import the installation media to.

engine :: Text

The name of the database engine to be used for this instance.

The list only includes supported DB engines that require an on-premises customer provided license.

Valid Values:

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

engineVersion :: Text

The version number of the database engine to use.

For a list of valid engine versions, call DescribeDBEngineVersions.

The following are the database engines and links to information about the major and minor versions. The list only includes DB engines that require an on-premises customer provided license.

Microsoft SQL Server

See Microsoft SQL Server Versions on Amazon RDS in the Amazon RDS User Guide.

engineInstallationMediaPath :: Text

The path to the installation medium for the specified DB engine.

Example: SQLServerISO/en_sql_server_2016_enterprise_x64_dvd_8701793.iso

oSInstallationMediaPath :: Text

The path to the installation medium for the operating system associated with the specified DB engine.

Example: WindowsISO/en_windows_server_2016_x64_dvd_9327751.iso

newImportInstallationMedia

Arguments

:: Text
$sel:customAvailabilityZoneId:ImportInstallationMedia'

-> Text
$sel:engine:ImportInstallationMedia'

-> Text
$sel:engineVersion:ImportInstallationMedia'

-> Text
$sel:engineInstallationMediaPath:ImportInstallationMedia'

-> Text
$sel:oSInstallationMediaPath:ImportInstallationMedia'

-> ImportInstallationMedia

Create a value of ImportInstallationMedia with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:customAvailabilityZoneId:ImportInstallationMedia', importInstallationMedia_customAvailabilityZoneId - The identifier of the custom Availability Zone (AZ) to import the installation media to.

$sel:engine:ImportInstallationMedia', importInstallationMedia_engine - The name of the database engine to be used for this instance.

The list only includes supported DB engines that require an on-premises customer provided license.

Valid Values:

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

$sel:engineVersion:ImportInstallationMedia', importInstallationMedia_engineVersion - The version number of the database engine to use.

For a list of valid engine versions, call DescribeDBEngineVersions.

The following are the database engines and links to information about the major and minor versions. The list only includes DB engines that require an on-premises customer provided license.

Microsoft SQL Server

See Microsoft SQL Server Versions on Amazon RDS in the Amazon RDS User Guide.

$sel:engineInstallationMediaPath:ImportInstallationMedia', importInstallationMedia_engineInstallationMediaPath - The path to the installation medium for the specified DB engine.

Example: SQLServerISO/en_sql_server_2016_enterprise_x64_dvd_8701793.iso

$sel:oSInstallationMediaPath:ImportInstallationMedia', importInstallationMedia_oSInstallationMediaPath - The path to the installation medium for the operating system associated with the specified DB engine.

Example: WindowsISO/en_windows_server_2016_x64_dvd_9327751.iso

Request Lenses

importInstallationMedia_customAvailabilityZoneId :: Lens' ImportInstallationMedia Text

The identifier of the custom Availability Zone (AZ) to import the installation media to.

importInstallationMedia_engine :: Lens' ImportInstallationMedia Text

The name of the database engine to be used for this instance.

The list only includes supported DB engines that require an on-premises customer provided license.

Valid Values:

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

importInstallationMedia_engineVersion :: Lens' ImportInstallationMedia Text

The version number of the database engine to use.

For a list of valid engine versions, call DescribeDBEngineVersions.

The following are the database engines and links to information about the major and minor versions. The list only includes DB engines that require an on-premises customer provided license.

Microsoft SQL Server

See Microsoft SQL Server Versions on Amazon RDS in the Amazon RDS User Guide.

importInstallationMedia_engineInstallationMediaPath :: Lens' ImportInstallationMedia Text

The path to the installation medium for the specified DB engine.

Example: SQLServerISO/en_sql_server_2016_enterprise_x64_dvd_8701793.iso

importInstallationMedia_oSInstallationMediaPath :: Lens' ImportInstallationMedia Text

The path to the installation medium for the operating system associated with the specified DB engine.

Example: WindowsISO/en_windows_server_2016_x64_dvd_9327751.iso

Destructuring the Response

data InstallationMedia

Contains the installation media for a DB engine that requires an on-premises customer provided license, such as Microsoft SQL Server.

See: newInstallationMedia smart constructor.

Constructors

InstallationMedia'

Fields

status :: Maybe Text

The status of the installation medium.

customAvailabilityZoneId :: Maybe Text

The custom Availability Zone (AZ) that contains the installation media.

installationMediaId :: Maybe Text

The installation medium ID.

engineVersion :: Maybe Text

The engine version of the DB engine.

failureCause :: Maybe InstallationMediaFailureCause

If an installation media failure occurred, the cause of the failure.

oSInstallationMediaPath :: Maybe Text

The path to the installation medium for the operating system associated with the DB engine.

engine :: Maybe Text

The DB engine.

engineInstallationMediaPath :: Maybe Text

The path to the installation medium for the DB engine.

newInstallationMedia :: InstallationMedia

Create a value of InstallationMedia with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:InstallationMedia', installationMedia_status - The status of the installation medium.

$sel:customAvailabilityZoneId:InstallationMedia', installationMedia_customAvailabilityZoneId - The custom Availability Zone (AZ) that contains the installation media.

$sel:installationMediaId:InstallationMedia', installationMedia_installationMediaId - The installation medium ID.

$sel:engineVersion:InstallationMedia', installationMedia_engineVersion - The engine version of the DB engine.

$sel:failureCause:InstallationMedia', installationMedia_failureCause - If an installation media failure occurred, the cause of the failure.

$sel:oSInstallationMediaPath:InstallationMedia', installationMedia_oSInstallationMediaPath - The path to the installation medium for the operating system associated with the DB engine.

$sel:engine:InstallationMedia', installationMedia_engine - The DB engine.

$sel:engineInstallationMediaPath:InstallationMedia', installationMedia_engineInstallationMediaPath - The path to the installation medium for the DB engine.

Response Lenses

installationMedia_status :: Lens' InstallationMedia (Maybe Text)

The status of the installation medium.

installationMedia_customAvailabilityZoneId :: Lens' InstallationMedia (Maybe Text)

The custom Availability Zone (AZ) that contains the installation media.

installationMedia_installationMediaId :: Lens' InstallationMedia (Maybe Text)

The installation medium ID.

installationMedia_engineVersion :: Lens' InstallationMedia (Maybe Text)

The engine version of the DB engine.

installationMedia_failureCause :: Lens' InstallationMedia (Maybe InstallationMediaFailureCause)

If an installation media failure occurred, the cause of the failure.

installationMedia_oSInstallationMediaPath :: Lens' InstallationMedia (Maybe Text)

The path to the installation medium for the operating system associated with the DB engine.

installationMedia_engine :: Lens' InstallationMedia (Maybe Text)

The DB engine.

installationMedia_engineInstallationMediaPath :: Lens' InstallationMedia (Maybe Text)

The path to the installation medium for the DB engine.