Network.AWS.RDS.ModifyCertificates

Description

Override the system-default Secure Sockets Layer/Transport Layer Security (SSL/TLS) certificate for Amazon RDS for new DB instances temporarily, or remove the override.

By using this operation, you can specify an RDS-approved SSL/TLS certificate for new DB instances that is different from the default certificate provided by RDS. You can also use this operation to remove the override, so that new DB instances use the default certificate provided by RDS.

You might need to override the default certificate in the following situations:

You already migrated your applications to support the latest certificate authority (CA) certificate, but the new CA certificate is not yet the RDS default CA certificate for the specified Amazon Web Services Region.

RDS has already moved to a new default CA certificate for the specified Amazon Web Services Region, but you are still in the process of supporting the new CA certificate. In this case, you temporarily need additional time to finish your application changes.

For more information about rotating your SSL/TLS certificate for RDS DB engines, see Rotating Your SSL/TLS Certificate in the Amazon RDS User Guide.

For more information about rotating your SSL/TLS certificate for Aurora DB engines, see Rotating Your SSL/TLS Certificate in the Amazon Aurora User Guide.

Creating a Request

data ModifyCertificates

See: newModifyCertificates smart constructor.

Constructors

ModifyCertificates'

Fields

certificateIdentifier :: Maybe Text

The new default certificate identifier to override the current one with.

To determine the valid values, use the describe-certificates CLI command or the DescribeCertificates API operation.

removeCustomerOverride :: Maybe Bool

A value that indicates whether to remove the override for the default certificate. If the override is removed, the default certificate is the system default.

newModifyCertificates :: ModifyCertificates

Create a value of ModifyCertificates with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificateIdentifier:ModifyCertificates', modifyCertificates_certificateIdentifier - The new default certificate identifier to override the current one with.

To determine the valid values, use the describe-certificates CLI command or the DescribeCertificates API operation.

$sel:removeCustomerOverride:ModifyCertificates', modifyCertificates_removeCustomerOverride - A value that indicates whether to remove the override for the default certificate. If the override is removed, the default certificate is the system default.

Request Lenses

modifyCertificates_certificateIdentifier :: Lens' ModifyCertificates (Maybe Text)

The new default certificate identifier to override the current one with.

To determine the valid values, use the describe-certificates CLI command or the DescribeCertificates API operation.

modifyCertificates_removeCustomerOverride :: Lens' ModifyCertificates (Maybe Bool)

A value that indicates whether to remove the override for the default certificate. If the override is removed, the default certificate is the system default.

Destructuring the Response

data ModifyCertificatesResponse

See: newModifyCertificatesResponse smart constructor.

Constructors

ModifyCertificatesResponse'

Fields

certificate :: Maybe Certificate

httpStatus :: Int

The response's http status code.

newModifyCertificatesResponse

Arguments

:: Int
$sel:httpStatus:ModifyCertificatesResponse'

-> ModifyCertificatesResponse

Create a value of ModifyCertificatesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:certificate:ModifyCertificatesResponse', modifyCertificatesResponse_certificate - Undocumented member.

$sel:httpStatus:ModifyCertificatesResponse', modifyCertificatesResponse_httpStatus - The response's http status code.

Response Lenses

modifyCertificatesResponse_certificate :: Lens' ModifyCertificatesResponse (Maybe Certificate)

Undocumented member.

modifyCertificatesResponse_httpStatus :: Lens' ModifyCertificatesResponse Int

The response's http status code.