Network.AWS.RDS.ModifyDBCluster

Description

Modify a setting for an Amazon Aurora DB cluster. You can change one or more database configuration parameters by specifying these parameters and the new values in the request. For more information on Amazon Aurora, see What Is Amazon Aurora? in the Amazon Aurora User Guide.

This action only applies to Aurora DB clusters.

Creating a Request

data ModifyDBCluster

See: newModifyDBCluster smart constructor.

Constructors

ModifyDBCluster'

Fields

deletionProtection :: Maybe Bool

A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

preferredBackupWindow :: Maybe Text

The daily time range during which automated backups are created if automated backups are enabled, using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region. To view the time blocks available, see Backup window in the Amazon Aurora User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

backupRetentionPeriod :: Maybe Int

The number of days for which automated backups are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35

enableIAMDatabaseAuthentication :: Maybe Bool

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information, see IAM Database Authentication in the Amazon Aurora User Guide.

dbInstanceParameterGroupName :: Maybe Text

The name of the DB parameter group to apply to all instances of the DB cluster.

When you apply a parameter group using the DBInstanceParameterGroupName parameter, the DB cluster isn't rebooted automatically. Also, parameter changes aren't applied during the next maintenance window but instead are applied immediately.

Default: The existing name setting

Constraints:

The DB parameter group must be in the same DB parameter group family as this DB cluster.

The DBInstanceParameterGroupName parameter is only valid in combination with the AllowMajorVersionUpgrade parameter.

optionGroupName :: Maybe Text

A value that indicates that the DB cluster should be associated with the specified option group. Changing this parameter doesn't result in an outage except in the following case, and the change is applied during the next maintenance window unless the ApplyImmediately is enabled for this request. If the parameter change results in an option group that enables OEM, this change can cause a brief (sub-second) period during which new connections are rejected but existing connections are not interrupted.

Permanent options can't be removed from an option group. The option group can't be removed from a DB cluster once it is associated with a DB cluster.

domain :: Maybe Text

The Active Directory directory ID to move the DB cluster to. Specify none to remove the cluster from its current domain. The domain must be created prior to this operation.

For more information, see Kerberos Authentication in the Amazon Aurora User Guide.

allowMajorVersionUpgrade :: Maybe Bool

A value that indicates whether major version upgrades are allowed.

Constraints: You must allow major version upgrades when specifying a value for the EngineVersion parameter that is a different major version than the DB cluster's current version.

scalingConfiguration :: Maybe ScalingConfiguration

The scaling properties of the DB cluster. You can only modify scaling properties for DB clusters in serverless DB engine mode.

masterUserPassword :: Maybe Text

The new password for the master database user. This password can contain any printable ASCII character except "/", """, or "@".

Constraints: Must contain from 8 to 41 characters.

enableHttpEndpoint :: Maybe Bool

A value that indicates whether to enable the HTTP endpoint for an Aurora Serverless DB cluster. By default, the HTTP endpoint is disabled.

When enabled, the HTTP endpoint provides a connectionless web service API for running SQL queries on the Aurora Serverless DB cluster. You can also query your database from inside the RDS console with the query editor.

For more information, see Using the Data API for Aurora Serverless in the Amazon Aurora User Guide.

vpcSecurityGroupIds :: Maybe [Text]

A list of VPC security groups that the DB cluster will belong to.

engineVersion :: Maybe Text

The version number of the database engine to which you want to upgrade. Changing this parameter results in an outage. The change is applied during the next maintenance window unless ApplyImmediately is enabled.

To list all of the available engine versions for aurora (for MySQL 5.6-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-mysql (for MySQL 5.7-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora-mysql --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-postgresql, use the following command:

aws rds describe-db-engine-versions --engine aurora-postgresql --query "DBEngineVersions[].EngineVersion"

enableGlobalWriteForwarding :: Maybe Bool

A value that indicates whether to enable this DB cluster to forward write operations to the primary cluster of an Aurora global database (GlobalCluster). By default, write operations are not allowed on Aurora DB clusters that are secondary clusters in an Aurora global database.

You can set this value only on Aurora DB clusters that are members of an Aurora global database. With this parameter enabled, a secondary cluster can forward writes to the current primary cluster and the resulting changes are replicated back to this cluster. For the primary DB cluster of an Aurora global database, this value is used immediately if the primary is demoted by the FailoverGlobalCluster API operation, but it does nothing until then.

preferredMaintenanceWindow :: Maybe Text

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred DB Cluster Maintenance Window in the Amazon Aurora User Guide.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

port :: Maybe Int

The port number on which the DB cluster accepts connections.

Constraints: Value must be 1150-65535

Default: The same port as the original DB cluster.

domainIAMRoleName :: Maybe Text

Specify the name of the IAM role to be used when making API calls to the Directory Service.

copyTagsToSnapshot :: Maybe Bool

A value that indicates whether to copy all tags from the DB cluster to snapshots of the DB cluster. The default is not to copy them.

newDBClusterIdentifier' :: Maybe Text

The new DB cluster identifier for the DB cluster when renaming a DB cluster. This value is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

The first character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: my-cluster2

cloudwatchLogsExportConfiguration :: Maybe CloudwatchLogsExportConfiguration

The configuration setting for the log types to be enabled for export to CloudWatch Logs for a specific DB cluster.

backtrackWindow :: Maybe Integer

The target backtrack window, in seconds. To disable backtracking, set this value to 0.

Currently, Backtrack is only supported for Aurora MySQL DB clusters.

Default: 0

Constraints:

If specified, this value must be set to a number from 0 to 259,200 (72 hours).

applyImmediately :: Maybe Bool

A value that indicates whether the modifications in this request and any pending modifications are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the DB cluster. If this parameter is disabled, changes to the DB cluster are applied during the next maintenance window.

The ApplyImmediately parameter only affects the EnableIAMDatabaseAuthentication, MasterUserPassword, and NewDBClusterIdentifier values. If the ApplyImmediately parameter is disabled, then changes to the EnableIAMDatabaseAuthentication, MasterUserPassword, and NewDBClusterIdentifier values are applied during the next maintenance window. All other changes are applied immediately, regardless of the value of the ApplyImmediately parameter.

By default, this parameter is disabled.

dbClusterParameterGroupName :: Maybe Text

The name of the DB cluster parameter group to use for the DB cluster.

dbClusterIdentifier :: Text

The DB cluster identifier for the cluster being modified. This parameter isn't case-sensitive.

Constraints: This identifier must match the identifier of an existing DB cluster.

newModifyDBCluster

Arguments

:: Text
$sel:dbClusterIdentifier:ModifyDBCluster'

-> ModifyDBCluster

Create a value of ModifyDBCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deletionProtection:ModifyDBCluster', modifyDBCluster_deletionProtection - A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

$sel:preferredBackupWindow:ModifyDBCluster', modifyDBCluster_preferredBackupWindow - The daily time range during which automated backups are created if automated backups are enabled, using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region. To view the time blocks available, see Backup window in the Amazon Aurora User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

$sel:backupRetentionPeriod:ModifyDBCluster', modifyDBCluster_backupRetentionPeriod - The number of days for which automated backups are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35

$sel:enableIAMDatabaseAuthentication:ModifyDBCluster', modifyDBCluster_enableIAMDatabaseAuthentication - A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information, see IAM Database Authentication in the Amazon Aurora User Guide.

$sel:dbInstanceParameterGroupName:ModifyDBCluster', modifyDBCluster_dbInstanceParameterGroupName - The name of the DB parameter group to apply to all instances of the DB cluster.

When you apply a parameter group using the DBInstanceParameterGroupName parameter, the DB cluster isn't rebooted automatically. Also, parameter changes aren't applied during the next maintenance window but instead are applied immediately.

Default: The existing name setting

Constraints:

The DB parameter group must be in the same DB parameter group family as this DB cluster.

The DBInstanceParameterGroupName parameter is only valid in combination with the AllowMajorVersionUpgrade parameter.

$sel:optionGroupName:ModifyDBCluster', modifyDBCluster_optionGroupName - A value that indicates that the DB cluster should be associated with the specified option group. Changing this parameter doesn't result in an outage except in the following case, and the change is applied during the next maintenance window unless the ApplyImmediately is enabled for this request. If the parameter change results in an option group that enables OEM, this change can cause a brief (sub-second) period during which new connections are rejected but existing connections are not interrupted.

Permanent options can't be removed from an option group. The option group can't be removed from a DB cluster once it is associated with a DB cluster.

$sel:domain:ModifyDBCluster', modifyDBCluster_domain - The Active Directory directory ID to move the DB cluster to. Specify none to remove the cluster from its current domain. The domain must be created prior to this operation.

For more information, see Kerberos Authentication in the Amazon Aurora User Guide.

$sel:allowMajorVersionUpgrade:ModifyDBCluster', modifyDBCluster_allowMajorVersionUpgrade - A value that indicates whether major version upgrades are allowed.

Constraints: You must allow major version upgrades when specifying a value for the EngineVersion parameter that is a different major version than the DB cluster's current version.

$sel:scalingConfiguration:ModifyDBCluster', modifyDBCluster_scalingConfiguration - The scaling properties of the DB cluster. You can only modify scaling properties for DB clusters in serverless DB engine mode.

$sel:masterUserPassword:ModifyDBCluster', modifyDBCluster_masterUserPassword - The new password for the master database user. This password can contain any printable ASCII character except "/", """, or "@".

Constraints: Must contain from 8 to 41 characters.

$sel:enableHttpEndpoint:ModifyDBCluster', modifyDBCluster_enableHttpEndpoint - A value that indicates whether to enable the HTTP endpoint for an Aurora Serverless DB cluster. By default, the HTTP endpoint is disabled.

When enabled, the HTTP endpoint provides a connectionless web service API for running SQL queries on the Aurora Serverless DB cluster. You can also query your database from inside the RDS console with the query editor.

For more information, see Using the Data API for Aurora Serverless in the Amazon Aurora User Guide.

$sel:vpcSecurityGroupIds:ModifyDBCluster', modifyDBCluster_vpcSecurityGroupIds - A list of VPC security groups that the DB cluster will belong to.

$sel:engineVersion:ModifyDBCluster', modifyDBCluster_engineVersion - The version number of the database engine to which you want to upgrade. Changing this parameter results in an outage. The change is applied during the next maintenance window unless ApplyImmediately is enabled.

To list all of the available engine versions for aurora (for MySQL 5.6-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-mysql (for MySQL 5.7-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora-mysql --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-postgresql, use the following command:

aws rds describe-db-engine-versions --engine aurora-postgresql --query "DBEngineVersions[].EngineVersion"

$sel:enableGlobalWriteForwarding:ModifyDBCluster', modifyDBCluster_enableGlobalWriteForwarding - A value that indicates whether to enable this DB cluster to forward write operations to the primary cluster of an Aurora global database (GlobalCluster). By default, write operations are not allowed on Aurora DB clusters that are secondary clusters in an Aurora global database.

You can set this value only on Aurora DB clusters that are members of an Aurora global database. With this parameter enabled, a secondary cluster can forward writes to the current primary cluster and the resulting changes are replicated back to this cluster. For the primary DB cluster of an Aurora global database, this value is used immediately if the primary is demoted by the FailoverGlobalCluster API operation, but it does nothing until then.

$sel:preferredMaintenanceWindow:ModifyDBCluster', modifyDBCluster_preferredMaintenanceWindow - The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred DB Cluster Maintenance Window in the Amazon Aurora User Guide.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

$sel:port:ModifyDBCluster', modifyDBCluster_port - The port number on which the DB cluster accepts connections.

Constraints: Value must be 1150-65535

Default: The same port as the original DB cluster.

$sel:domainIAMRoleName:ModifyDBCluster', modifyDBCluster_domainIAMRoleName - Specify the name of the IAM role to be used when making API calls to the Directory Service.

$sel:copyTagsToSnapshot:ModifyDBCluster', modifyDBCluster_copyTagsToSnapshot - A value that indicates whether to copy all tags from the DB cluster to snapshots of the DB cluster. The default is not to copy them.

$sel:newDBClusterIdentifier':ModifyDBCluster', modifyDBCluster_newDBClusterIdentifier - The new DB cluster identifier for the DB cluster when renaming a DB cluster. This value is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

The first character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: my-cluster2

$sel:cloudwatchLogsExportConfiguration:ModifyDBCluster', modifyDBCluster_cloudwatchLogsExportConfiguration - The configuration setting for the log types to be enabled for export to CloudWatch Logs for a specific DB cluster.

$sel:backtrackWindow:ModifyDBCluster', modifyDBCluster_backtrackWindow - The target backtrack window, in seconds. To disable backtracking, set this value to 0.

Currently, Backtrack is only supported for Aurora MySQL DB clusters.

Default: 0

Constraints:

If specified, this value must be set to a number from 0 to 259,200 (72 hours).

$sel:applyImmediately:ModifyDBCluster', modifyDBCluster_applyImmediately - A value that indicates whether the modifications in this request and any pending modifications are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the DB cluster. If this parameter is disabled, changes to the DB cluster are applied during the next maintenance window.

The ApplyImmediately parameter only affects the EnableIAMDatabaseAuthentication, MasterUserPassword, and NewDBClusterIdentifier values. If the ApplyImmediately parameter is disabled, then changes to the EnableIAMDatabaseAuthentication, MasterUserPassword, and NewDBClusterIdentifier values are applied during the next maintenance window. All other changes are applied immediately, regardless of the value of the ApplyImmediately parameter.

By default, this parameter is disabled.

$sel:dbClusterParameterGroupName:ModifyDBCluster', modifyDBCluster_dbClusterParameterGroupName - The name of the DB cluster parameter group to use for the DB cluster.

$sel:dbClusterIdentifier:ModifyDBCluster', modifyDBCluster_dbClusterIdentifier - The DB cluster identifier for the cluster being modified. This parameter isn't case-sensitive.

Constraints: This identifier must match the identifier of an existing DB cluster.

Request Lenses

modifyDBCluster_deletionProtection :: Lens' ModifyDBCluster (Maybe Bool)

A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

modifyDBCluster_preferredBackupWindow :: Lens' ModifyDBCluster (Maybe Text)

The daily time range during which automated backups are created if automated backups are enabled, using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region. To view the time blocks available, see Backup window in the Amazon Aurora User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

modifyDBCluster_backupRetentionPeriod :: Lens' ModifyDBCluster (Maybe Int)

The number of days for which automated backups are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35

modifyDBCluster_enableIAMDatabaseAuthentication :: Lens' ModifyDBCluster (Maybe Bool)

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information, see IAM Database Authentication in the Amazon Aurora User Guide.

modifyDBCluster_dbInstanceParameterGroupName :: Lens' ModifyDBCluster (Maybe Text)

The name of the DB parameter group to apply to all instances of the DB cluster.

When you apply a parameter group using the DBInstanceParameterGroupName parameter, the DB cluster isn't rebooted automatically. Also, parameter changes aren't applied during the next maintenance window but instead are applied immediately.

Default: The existing name setting

Constraints:

The DB parameter group must be in the same DB parameter group family as this DB cluster.

The DBInstanceParameterGroupName parameter is only valid in combination with the AllowMajorVersionUpgrade parameter.

modifyDBCluster_optionGroupName :: Lens' ModifyDBCluster (Maybe Text)

A value that indicates that the DB cluster should be associated with the specified option group. Changing this parameter doesn't result in an outage except in the following case, and the change is applied during the next maintenance window unless the ApplyImmediately is enabled for this request. If the parameter change results in an option group that enables OEM, this change can cause a brief (sub-second) period during which new connections are rejected but existing connections are not interrupted.

Permanent options can't be removed from an option group. The option group can't be removed from a DB cluster once it is associated with a DB cluster.

modifyDBCluster_domain :: Lens' ModifyDBCluster (Maybe Text)

The Active Directory directory ID to move the DB cluster to. Specify none to remove the cluster from its current domain. The domain must be created prior to this operation.

For more information, see Kerberos Authentication in the Amazon Aurora User Guide.

modifyDBCluster_allowMajorVersionUpgrade :: Lens' ModifyDBCluster (Maybe Bool)

A value that indicates whether major version upgrades are allowed.

Constraints: You must allow major version upgrades when specifying a value for the EngineVersion parameter that is a different major version than the DB cluster's current version.

modifyDBCluster_scalingConfiguration :: Lens' ModifyDBCluster (Maybe ScalingConfiguration)

The scaling properties of the DB cluster. You can only modify scaling properties for DB clusters in serverless DB engine mode.

modifyDBCluster_masterUserPassword :: Lens' ModifyDBCluster (Maybe Text)

The new password for the master database user. This password can contain any printable ASCII character except "/", """, or "@".

Constraints: Must contain from 8 to 41 characters.

modifyDBCluster_enableHttpEndpoint :: Lens' ModifyDBCluster (Maybe Bool)

A value that indicates whether to enable the HTTP endpoint for an Aurora Serverless DB cluster. By default, the HTTP endpoint is disabled.

When enabled, the HTTP endpoint provides a connectionless web service API for running SQL queries on the Aurora Serverless DB cluster. You can also query your database from inside the RDS console with the query editor.

For more information, see Using the Data API for Aurora Serverless in the Amazon Aurora User Guide.

modifyDBCluster_vpcSecurityGroupIds :: Lens' ModifyDBCluster (Maybe [Text])

A list of VPC security groups that the DB cluster will belong to.

modifyDBCluster_engineVersion :: Lens' ModifyDBCluster (Maybe Text)

The version number of the database engine to which you want to upgrade. Changing this parameter results in an outage. The change is applied during the next maintenance window unless ApplyImmediately is enabled.

To list all of the available engine versions for aurora (for MySQL 5.6-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-mysql (for MySQL 5.7-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora-mysql --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-postgresql, use the following command:

aws rds describe-db-engine-versions --engine aurora-postgresql --query "DBEngineVersions[].EngineVersion"

modifyDBCluster_enableGlobalWriteForwarding :: Lens' ModifyDBCluster (Maybe Bool)

A value that indicates whether to enable this DB cluster to forward write operations to the primary cluster of an Aurora global database (GlobalCluster). By default, write operations are not allowed on Aurora DB clusters that are secondary clusters in an Aurora global database.

You can set this value only on Aurora DB clusters that are members of an Aurora global database. With this parameter enabled, a secondary cluster can forward writes to the current primary cluster and the resulting changes are replicated back to this cluster. For the primary DB cluster of an Aurora global database, this value is used immediately if the primary is demoted by the FailoverGlobalCluster API operation, but it does nothing until then.

modifyDBCluster_preferredMaintenanceWindow :: Lens' ModifyDBCluster (Maybe Text)

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred DB Cluster Maintenance Window in the Amazon Aurora User Guide.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

modifyDBCluster_port :: Lens' ModifyDBCluster (Maybe Int)

The port number on which the DB cluster accepts connections.

Constraints: Value must be 1150-65535

Default: The same port as the original DB cluster.

modifyDBCluster_domainIAMRoleName :: Lens' ModifyDBCluster (Maybe Text)

Specify the name of the IAM role to be used when making API calls to the Directory Service.

modifyDBCluster_copyTagsToSnapshot :: Lens' ModifyDBCluster (Maybe Bool)

A value that indicates whether to copy all tags from the DB cluster to snapshots of the DB cluster. The default is not to copy them.

modifyDBCluster_newDBClusterIdentifier :: Lens' ModifyDBCluster (Maybe Text)

The new DB cluster identifier for the DB cluster when renaming a DB cluster. This value is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

The first character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: my-cluster2

modifyDBCluster_cloudwatchLogsExportConfiguration :: Lens' ModifyDBCluster (Maybe CloudwatchLogsExportConfiguration)

The configuration setting for the log types to be enabled for export to CloudWatch Logs for a specific DB cluster.

modifyDBCluster_backtrackWindow :: Lens' ModifyDBCluster (Maybe Integer)

The target backtrack window, in seconds. To disable backtracking, set this value to 0.

Currently, Backtrack is only supported for Aurora MySQL DB clusters.

Default: 0

Constraints:

If specified, this value must be set to a number from 0 to 259,200 (72 hours).

modifyDBCluster_applyImmediately :: Lens' ModifyDBCluster (Maybe Bool)

A value that indicates whether the modifications in this request and any pending modifications are asynchronously applied as soon as possible, regardless of the PreferredMaintenanceWindow setting for the DB cluster. If this parameter is disabled, changes to the DB cluster are applied during the next maintenance window.

The ApplyImmediately parameter only affects the EnableIAMDatabaseAuthentication, MasterUserPassword, and NewDBClusterIdentifier values. If the ApplyImmediately parameter is disabled, then changes to the EnableIAMDatabaseAuthentication, MasterUserPassword, and NewDBClusterIdentifier values are applied during the next maintenance window. All other changes are applied immediately, regardless of the value of the ApplyImmediately parameter.

By default, this parameter is disabled.

modifyDBCluster_dbClusterParameterGroupName :: Lens' ModifyDBCluster (Maybe Text)

The name of the DB cluster parameter group to use for the DB cluster.

modifyDBCluster_dbClusterIdentifier :: Lens' ModifyDBCluster Text

The DB cluster identifier for the cluster being modified. This parameter isn't case-sensitive.

Constraints: This identifier must match the identifier of an existing DB cluster.

Destructuring the Response

data ModifyDBClusterResponse

See: newModifyDBClusterResponse smart constructor.

Constructors

ModifyDBClusterResponse'

Fields

dbCluster :: Maybe DBCluster

httpStatus :: Int

The response's http status code.

newModifyDBClusterResponse

Arguments

:: Int
$sel:httpStatus:ModifyDBClusterResponse'

-> ModifyDBClusterResponse

Create a value of ModifyDBClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbCluster:ModifyDBClusterResponse', modifyDBClusterResponse_dbCluster - Undocumented member.

$sel:httpStatus:ModifyDBClusterResponse', modifyDBClusterResponse_httpStatus - The response's http status code.

Response Lenses

modifyDBClusterResponse_dbCluster :: Lens' ModifyDBClusterResponse (Maybe DBCluster)

Undocumented member.

modifyDBClusterResponse_httpStatus :: Lens' ModifyDBClusterResponse Int

The response's http status code.