Network.AWS.RDS.ModifyDBParameterGroup

Description

Modifies the parameters of a DB parameter group. To modify more than one parameter, submit a list of the following: ParameterName, ParameterValue, and ApplyMethod. A maximum of 20 parameters can be modified in a single request.

After you modify a DB parameter group, you should wait at least 5 minutes before creating your first DB instance that uses that DB parameter group as the default parameter group. This allows Amazon RDS to fully complete the modify action before the parameter group is used as the default for a new DB instance. This is especially important for parameters that are critical when creating the default database for a DB instance, such as the character set for the default database defined by the character_set_database parameter. You can use the /Parameter Groups/ option of the Amazon RDS console or the DescribeDBParameters command to verify that your DB parameter group has been created or modified.

Creating a Request

data ModifyDBParameterGroup

See: newModifyDBParameterGroup smart constructor.

Constructors

ModifyDBParameterGroup'

Fields

dbParameterGroupName :: Text

The name of the DB parameter group.

Constraints:

If supplied, must match the name of an existing DBParameterGroup.

parameters :: [Parameter]

An array of parameter names, values, and the application methods for the parameter update. At least one parameter name, value, and application method method must be supplied; later arguments are optional. A maximum of 20 parameters can be modified in a single request.

Valid Values (for the application method): immediate | pending-reboot

You can use the immediate value with dynamic parameters only. You can use the pending-reboot value for both dynamic and static parameters.

When the application method is immediate, changes to dynamic parameters are applied immediately to the DB instances associated with the parameter group. When the application method is pending-reboot, changes to dynamic and static parameters are applied after a reboot without failover to the DB instances associated with the parameter group.

newModifyDBParameterGroup

Arguments

:: Text
$sel:dbParameterGroupName:ModifyDBParameterGroup'

-> ModifyDBParameterGroup

Create a value of ModifyDBParameterGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbParameterGroupName:ModifyDBParameterGroup', modifyDBParameterGroup_dbParameterGroupName - The name of the DB parameter group.

Constraints:

If supplied, must match the name of an existing DBParameterGroup.

$sel:parameters:ModifyDBParameterGroup', modifyDBParameterGroup_parameters - An array of parameter names, values, and the application methods for the parameter update. At least one parameter name, value, and application method method must be supplied; later arguments are optional. A maximum of 20 parameters can be modified in a single request.

Valid Values (for the application method): immediate | pending-reboot

You can use the immediate value with dynamic parameters only. You can use the pending-reboot value for both dynamic and static parameters.

When the application method is immediate, changes to dynamic parameters are applied immediately to the DB instances associated with the parameter group. When the application method is pending-reboot, changes to dynamic and static parameters are applied after a reboot without failover to the DB instances associated with the parameter group.

Request Lenses

modifyDBParameterGroup_dbParameterGroupName :: Lens' ModifyDBParameterGroup Text

The name of the DB parameter group.

Constraints:

If supplied, must match the name of an existing DBParameterGroup.

modifyDBParameterGroup_parameters :: Lens' ModifyDBParameterGroup [Parameter]

An array of parameter names, values, and the application methods for the parameter update. At least one parameter name, value, and application method method must be supplied; later arguments are optional. A maximum of 20 parameters can be modified in a single request.

Valid Values (for the application method): immediate | pending-reboot

You can use the immediate value with dynamic parameters only. You can use the pending-reboot value for both dynamic and static parameters.

When the application method is immediate, changes to dynamic parameters are applied immediately to the DB instances associated with the parameter group. When the application method is pending-reboot, changes to dynamic and static parameters are applied after a reboot without failover to the DB instances associated with the parameter group.

Destructuring the Response

data DBParameterGroupNameMessage

Contains the result of a successful invocation of the ModifyDBParameterGroup or ResetDBParameterGroup action.

See: newDBParameterGroupNameMessage smart constructor.

Constructors

DBParameterGroupNameMessage'

Fields

dbParameterGroupName :: Maybe Text

The name of the DB parameter group.

newDBParameterGroupNameMessage :: DBParameterGroupNameMessage

Create a value of DBParameterGroupNameMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbParameterGroupName:DBParameterGroupNameMessage', dbParameterGroupNameMessage_dbParameterGroupName - The name of the DB parameter group.

Response Lenses

dbParameterGroupNameMessage_dbParameterGroupName :: Lens' DBParameterGroupNameMessage (Maybe Text)

The name of the DB parameter group.