Network.AWS.RDS.ModifyDBProxy

Description

Changes the settings for an existing DB proxy.

Creating a Request

data ModifyDBProxy

See: newModifyDBProxy smart constructor.

Constructors

ModifyDBProxy'

Fields

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role that the proxy uses to access secrets in Amazon Web Services Secrets Manager.

newDBProxyName' :: Maybe Text

The new identifier for the DBProxy. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens; it can't end with a hyphen or contain two consecutive hyphens.

idleClientTimeout :: Maybe Int

The number of seconds that a connection to the proxy can be inactive before the proxy disconnects it. You can set this value higher or lower than the connection timeout limit for the associated database.

auth :: Maybe [UserAuthConfig]

The new authentication settings for the DBProxy.

debugLogging :: Maybe Bool

Whether the proxy includes detailed information about SQL statements in its logs. This information helps you to debug issues involving SQL behavior or the performance and scalability of the proxy connections. The debug information includes the text of SQL statements that you submit through the proxy. Thus, only enable this setting when needed for debugging, and only when you have security measures in place to safeguard any sensitive information that appears in the logs.

securityGroups :: Maybe [Text]

The new list of security groups for the DBProxy.

requireTLS :: Maybe Bool

Whether Transport Layer Security (TLS) encryption is required for connections to the proxy. By enabling this setting, you can enforce encrypted TLS connections to the proxy, even if the associated database doesn't use TLS.

dbProxyName :: Text

The identifier for the DBProxy to modify.

newModifyDBProxy

Arguments

:: Text
$sel:dbProxyName:ModifyDBProxy'

-> ModifyDBProxy

Create a value of ModifyDBProxy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:roleArn:ModifyDBProxy', modifyDBProxy_roleArn - The Amazon Resource Name (ARN) of the IAM role that the proxy uses to access secrets in Amazon Web Services Secrets Manager.

$sel:newDBProxyName':ModifyDBProxy', modifyDBProxy_newDBProxyName - The new identifier for the DBProxy. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens; it can't end with a hyphen or contain two consecutive hyphens.

$sel:idleClientTimeout:ModifyDBProxy', modifyDBProxy_idleClientTimeout - The number of seconds that a connection to the proxy can be inactive before the proxy disconnects it. You can set this value higher or lower than the connection timeout limit for the associated database.

$sel:auth:ModifyDBProxy', modifyDBProxy_auth - The new authentication settings for the DBProxy.

$sel:debugLogging:ModifyDBProxy', modifyDBProxy_debugLogging - Whether the proxy includes detailed information about SQL statements in its logs. This information helps you to debug issues involving SQL behavior or the performance and scalability of the proxy connections. The debug information includes the text of SQL statements that you submit through the proxy. Thus, only enable this setting when needed for debugging, and only when you have security measures in place to safeguard any sensitive information that appears in the logs.

$sel:securityGroups:ModifyDBProxy', modifyDBProxy_securityGroups - The new list of security groups for the DBProxy.

$sel:requireTLS:ModifyDBProxy', modifyDBProxy_requireTLS - Whether Transport Layer Security (TLS) encryption is required for connections to the proxy. By enabling this setting, you can enforce encrypted TLS connections to the proxy, even if the associated database doesn't use TLS.

$sel:dbProxyName:ModifyDBProxy', modifyDBProxy_dbProxyName - The identifier for the DBProxy to modify.

Request Lenses

modifyDBProxy_roleArn :: Lens' ModifyDBProxy (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role that the proxy uses to access secrets in Amazon Web Services Secrets Manager.

modifyDBProxy_newDBProxyName :: Lens' ModifyDBProxy (Maybe Text)

The new identifier for the DBProxy. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens; it can't end with a hyphen or contain two consecutive hyphens.

modifyDBProxy_idleClientTimeout :: Lens' ModifyDBProxy (Maybe Int)

The number of seconds that a connection to the proxy can be inactive before the proxy disconnects it. You can set this value higher or lower than the connection timeout limit for the associated database.

modifyDBProxy_auth :: Lens' ModifyDBProxy (Maybe [UserAuthConfig])

The new authentication settings for the DBProxy.

modifyDBProxy_debugLogging :: Lens' ModifyDBProxy (Maybe Bool)

Whether the proxy includes detailed information about SQL statements in its logs. This information helps you to debug issues involving SQL behavior or the performance and scalability of the proxy connections. The debug information includes the text of SQL statements that you submit through the proxy. Thus, only enable this setting when needed for debugging, and only when you have security measures in place to safeguard any sensitive information that appears in the logs.

modifyDBProxy_securityGroups :: Lens' ModifyDBProxy (Maybe [Text])

The new list of security groups for the DBProxy.

modifyDBProxy_requireTLS :: Lens' ModifyDBProxy (Maybe Bool)

Whether Transport Layer Security (TLS) encryption is required for connections to the proxy. By enabling this setting, you can enforce encrypted TLS connections to the proxy, even if the associated database doesn't use TLS.

modifyDBProxy_dbProxyName :: Lens' ModifyDBProxy Text

The identifier for the DBProxy to modify.

Destructuring the Response

data ModifyDBProxyResponse

See: newModifyDBProxyResponse smart constructor.

Constructors

ModifyDBProxyResponse'

Fields

dbProxy :: Maybe DBProxy

The DBProxy object representing the new settings for the proxy.

httpStatus :: Int

The response's http status code.

newModifyDBProxyResponse

Arguments

:: Int
$sel:httpStatus:ModifyDBProxyResponse'

-> ModifyDBProxyResponse

Create a value of ModifyDBProxyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbProxy:ModifyDBProxyResponse', modifyDBProxyResponse_dbProxy - The DBProxy object representing the new settings for the proxy.

$sel:httpStatus:ModifyDBProxyResponse', modifyDBProxyResponse_httpStatus - The response's http status code.

Response Lenses

modifyDBProxyResponse_dbProxy :: Lens' ModifyDBProxyResponse (Maybe DBProxy)

The DBProxy object representing the new settings for the proxy.

modifyDBProxyResponse_httpStatus :: Lens' ModifyDBProxyResponse Int

The response's http status code.