Network.AWS.RDS.ModifyDBProxyEndpoint

Description

Changes the settings for an existing DB proxy endpoint.

Creating a Request

data ModifyDBProxyEndpoint

See: newModifyDBProxyEndpoint smart constructor.

Constructors

ModifyDBProxyEndpoint'

Fields

vpcSecurityGroupIds :: Maybe [Text]

The VPC security group IDs for the DB proxy endpoint. When the DB proxy endpoint uses a different VPC than the original proxy, you also specify a different set of security group IDs than for the original proxy.

newDBProxyEndpointName' :: Maybe Text

The new identifier for the DBProxyEndpoint. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens; it can't end with a hyphen or contain two consecutive hyphens.

dbProxyEndpointName :: Text

The name of the DB proxy sociated with the DB proxy endpoint that you want to modify.

newModifyDBProxyEndpoint

Arguments

:: Text
$sel:dbProxyEndpointName:ModifyDBProxyEndpoint'

-> ModifyDBProxyEndpoint

Create a value of ModifyDBProxyEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcSecurityGroupIds:ModifyDBProxyEndpoint', modifyDBProxyEndpoint_vpcSecurityGroupIds - The VPC security group IDs for the DB proxy endpoint. When the DB proxy endpoint uses a different VPC than the original proxy, you also specify a different set of security group IDs than for the original proxy.

$sel:newDBProxyEndpointName':ModifyDBProxyEndpoint', modifyDBProxyEndpoint_newDBProxyEndpointName - The new identifier for the DBProxyEndpoint. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens; it can't end with a hyphen or contain two consecutive hyphens.

$sel:dbProxyEndpointName:ModifyDBProxyEndpoint', modifyDBProxyEndpoint_dbProxyEndpointName - The name of the DB proxy sociated with the DB proxy endpoint that you want to modify.

Request Lenses

modifyDBProxyEndpoint_vpcSecurityGroupIds :: Lens' ModifyDBProxyEndpoint (Maybe [Text])

The VPC security group IDs for the DB proxy endpoint. When the DB proxy endpoint uses a different VPC than the original proxy, you also specify a different set of security group IDs than for the original proxy.

modifyDBProxyEndpoint_newDBProxyEndpointName :: Lens' ModifyDBProxyEndpoint (Maybe Text)

The new identifier for the DBProxyEndpoint. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens; it can't end with a hyphen or contain two consecutive hyphens.

modifyDBProxyEndpoint_dbProxyEndpointName :: Lens' ModifyDBProxyEndpoint Text

The name of the DB proxy sociated with the DB proxy endpoint that you want to modify.

Destructuring the Response

data ModifyDBProxyEndpointResponse

See: newModifyDBProxyEndpointResponse smart constructor.

Constructors

ModifyDBProxyEndpointResponse'

Fields

dbProxyEndpoint :: Maybe DBProxyEndpoint

The DBProxyEndpoint object representing the new settings for the DB proxy endpoint.

httpStatus :: Int

The response's http status code.

newModifyDBProxyEndpointResponse

Arguments

:: Int
$sel:httpStatus:ModifyDBProxyEndpointResponse'

-> ModifyDBProxyEndpointResponse

Create a value of ModifyDBProxyEndpointResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbProxyEndpoint:ModifyDBProxyEndpointResponse', modifyDBProxyEndpointResponse_dbProxyEndpoint - The DBProxyEndpoint object representing the new settings for the DB proxy endpoint.

$sel:httpStatus:ModifyDBProxyEndpointResponse', modifyDBProxyEndpointResponse_httpStatus - The response's http status code.

Response Lenses

modifyDBProxyEndpointResponse_dbProxyEndpoint :: Lens' ModifyDBProxyEndpointResponse (Maybe DBProxyEndpoint)

The DBProxyEndpoint object representing the new settings for the DB proxy endpoint.

modifyDBProxyEndpointResponse_httpStatus :: Lens' ModifyDBProxyEndpointResponse Int

The response's http status code.