Network.AWS.RDS.ModifyDBProxyTargetGroup

Description

Modifies the properties of a DBProxyTargetGroup.

Creating a Request

data ModifyDBProxyTargetGroup

See: newModifyDBProxyTargetGroup smart constructor.

Constructors

ModifyDBProxyTargetGroup'

Fields

connectionPoolConfig :: Maybe ConnectionPoolConfiguration

The settings that determine the size and behavior of the connection pool for the target group.

newName' :: Maybe Text

The new name for the modified DBProxyTarget. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens; it can't end with a hyphen or contain two consecutive hyphens.

targetGroupName :: Text

The name of the new target group to assign to the proxy.

dbProxyName :: Text

The name of the new proxy to which to assign the target group.

newModifyDBProxyTargetGroup

Arguments

:: Text
$sel:targetGroupName:ModifyDBProxyTargetGroup'

-> Text
$sel:dbProxyName:ModifyDBProxyTargetGroup'

-> ModifyDBProxyTargetGroup

Create a value of ModifyDBProxyTargetGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:connectionPoolConfig:ModifyDBProxyTargetGroup', modifyDBProxyTargetGroup_connectionPoolConfig - The settings that determine the size and behavior of the connection pool for the target group.

$sel:newName':ModifyDBProxyTargetGroup', modifyDBProxyTargetGroup_newName - The new name for the modified DBProxyTarget. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens; it can't end with a hyphen or contain two consecutive hyphens.

$sel:targetGroupName:ModifyDBProxyTargetGroup', modifyDBProxyTargetGroup_targetGroupName - The name of the new target group to assign to the proxy.

$sel:dbProxyName:ModifyDBProxyTargetGroup', modifyDBProxyTargetGroup_dbProxyName - The name of the new proxy to which to assign the target group.

Request Lenses

modifyDBProxyTargetGroup_connectionPoolConfig :: Lens' ModifyDBProxyTargetGroup (Maybe ConnectionPoolConfiguration)

The settings that determine the size and behavior of the connection pool for the target group.

modifyDBProxyTargetGroup_newName :: Lens' ModifyDBProxyTargetGroup (Maybe Text)

The new name for the modified DBProxyTarget. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens; it can't end with a hyphen or contain two consecutive hyphens.

modifyDBProxyTargetGroup_targetGroupName :: Lens' ModifyDBProxyTargetGroup Text

The name of the new target group to assign to the proxy.

modifyDBProxyTargetGroup_dbProxyName :: Lens' ModifyDBProxyTargetGroup Text

The name of the new proxy to which to assign the target group.

Destructuring the Response

data ModifyDBProxyTargetGroupResponse

See: newModifyDBProxyTargetGroupResponse smart constructor.

Constructors

ModifyDBProxyTargetGroupResponse'

Fields

dbProxyTargetGroup :: Maybe DBProxyTargetGroup

The settings of the modified DBProxyTarget.

httpStatus :: Int

The response's http status code.

newModifyDBProxyTargetGroupResponse

Arguments

:: Int
$sel:httpStatus:ModifyDBProxyTargetGroupResponse'

-> ModifyDBProxyTargetGroupResponse

Create a value of ModifyDBProxyTargetGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbProxyTargetGroup:ModifyDBProxyTargetGroupResponse', modifyDBProxyTargetGroupResponse_dbProxyTargetGroup - The settings of the modified DBProxyTarget.

$sel:httpStatus:ModifyDBProxyTargetGroupResponse', modifyDBProxyTargetGroupResponse_httpStatus - The response's http status code.

Response Lenses

modifyDBProxyTargetGroupResponse_dbProxyTargetGroup :: Lens' ModifyDBProxyTargetGroupResponse (Maybe DBProxyTargetGroup)

The settings of the modified DBProxyTarget.

modifyDBProxyTargetGroupResponse_httpStatus :: Lens' ModifyDBProxyTargetGroupResponse Int

The response's http status code.