Network.AWS.RDS.ModifyDBSnapshot

Description

Updates a manual DB snapshot with a new engine version. The snapshot can be encrypted or unencrypted, but not shared or public.

Amazon RDS supports upgrading DB snapshots for MySQL, Oracle, and PostgreSQL.

Creating a Request

data ModifyDBSnapshot

See: newModifyDBSnapshot smart constructor.

Constructors

ModifyDBSnapshot'

Fields

optionGroupName :: Maybe Text

The option group to identify with the upgraded DB snapshot.

You can specify this parameter when you upgrade an Oracle DB snapshot. The same option group considerations apply when upgrading a DB snapshot as when upgrading a DB instance. For more information, see Option group considerations in the Amazon RDS User Guide.

engineVersion :: Maybe Text

The engine version to upgrade the DB snapshot to.

The following are the database engines and engine versions that are available when you upgrade a DB snapshot.

MySQL

5.5.46 (supported for 5.1 DB snapshots)

Oracle

12.1.0.2.v8 (supported for 12.1.0.1 DB snapshots)

11.2.0.4.v12 (supported for 11.2.0.2 DB snapshots)

11.2.0.4.v11 (supported for 11.2.0.3 DB snapshots)

PostgreSQL

For the list of engine versions that are available for upgrading a DB snapshot, see Upgrading the PostgreSQL DB Engine for Amazon RDS.

dbSnapshotIdentifier :: Text

The identifier of the DB snapshot to modify.

newModifyDBSnapshot

Arguments

:: Text
$sel:dbSnapshotIdentifier:ModifyDBSnapshot'

-> ModifyDBSnapshot

Create a value of ModifyDBSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:optionGroupName:ModifyDBSnapshot', modifyDBSnapshot_optionGroupName - The option group to identify with the upgraded DB snapshot.

You can specify this parameter when you upgrade an Oracle DB snapshot. The same option group considerations apply when upgrading a DB snapshot as when upgrading a DB instance. For more information, see Option group considerations in the Amazon RDS User Guide.

$sel:engineVersion:ModifyDBSnapshot', modifyDBSnapshot_engineVersion - The engine version to upgrade the DB snapshot to.

The following are the database engines and engine versions that are available when you upgrade a DB snapshot.

MySQL

5.5.46 (supported for 5.1 DB snapshots)

Oracle

12.1.0.2.v8 (supported for 12.1.0.1 DB snapshots)

11.2.0.4.v12 (supported for 11.2.0.2 DB snapshots)

11.2.0.4.v11 (supported for 11.2.0.3 DB snapshots)

PostgreSQL

For the list of engine versions that are available for upgrading a DB snapshot, see Upgrading the PostgreSQL DB Engine for Amazon RDS.

$sel:dbSnapshotIdentifier:ModifyDBSnapshot', modifyDBSnapshot_dbSnapshotIdentifier - The identifier of the DB snapshot to modify.

Request Lenses

modifyDBSnapshot_optionGroupName :: Lens' ModifyDBSnapshot (Maybe Text)

The option group to identify with the upgraded DB snapshot.

You can specify this parameter when you upgrade an Oracle DB snapshot. The same option group considerations apply when upgrading a DB snapshot as when upgrading a DB instance. For more information, see Option group considerations in the Amazon RDS User Guide.

modifyDBSnapshot_engineVersion :: Lens' ModifyDBSnapshot (Maybe Text)

The engine version to upgrade the DB snapshot to.

The following are the database engines and engine versions that are available when you upgrade a DB snapshot.

MySQL

5.5.46 (supported for 5.1 DB snapshots)

Oracle

12.1.0.2.v8 (supported for 12.1.0.1 DB snapshots)

11.2.0.4.v12 (supported for 11.2.0.2 DB snapshots)

11.2.0.4.v11 (supported for 11.2.0.3 DB snapshots)

PostgreSQL

For the list of engine versions that are available for upgrading a DB snapshot, see Upgrading the PostgreSQL DB Engine for Amazon RDS.

modifyDBSnapshot_dbSnapshotIdentifier :: Lens' ModifyDBSnapshot Text

The identifier of the DB snapshot to modify.

Destructuring the Response

data ModifyDBSnapshotResponse

See: newModifyDBSnapshotResponse smart constructor.

Constructors

ModifyDBSnapshotResponse'

Fields

dbSnapshot :: Maybe DBSnapshot

httpStatus :: Int

The response's http status code.

newModifyDBSnapshotResponse

Arguments

:: Int
$sel:httpStatus:ModifyDBSnapshotResponse'

-> ModifyDBSnapshotResponse

Create a value of ModifyDBSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbSnapshot:ModifyDBSnapshotResponse', modifyDBSnapshotResponse_dbSnapshot - Undocumented member.

$sel:httpStatus:ModifyDBSnapshotResponse', modifyDBSnapshotResponse_httpStatus - The response's http status code.

Response Lenses

modifyDBSnapshotResponse_dbSnapshot :: Lens' ModifyDBSnapshotResponse (Maybe DBSnapshot)

Undocumented member.

modifyDBSnapshotResponse_httpStatus :: Lens' ModifyDBSnapshotResponse Int

The response's http status code.