Network.AWS.RDS.ModifyOptionGroup

Description

Modifies an existing option group.

Creating a Request

data ModifyOptionGroup

See: newModifyOptionGroup smart constructor.

Constructors

ModifyOptionGroup'

Fields

optionsToInclude :: Maybe [OptionConfiguration]

Options in this list are added to the option group or, if already present, the specified configuration is used to update the existing configuration.

optionsToRemove :: Maybe [Text]

Options in this list are removed from the option group.

applyImmediately :: Maybe Bool

A value that indicates whether to apply the change immediately or during the next maintenance window for each instance associated with the option group.

optionGroupName :: Text

The name of the option group to be modified.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, can't be removed from an option group, and that option group can't be removed from a DB instance once it is associated with a DB instance

newModifyOptionGroup

Arguments

:: Text
$sel:optionGroupName:ModifyOptionGroup'

-> ModifyOptionGroup

Create a value of ModifyOptionGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:optionsToInclude:ModifyOptionGroup', modifyOptionGroup_optionsToInclude - Options in this list are added to the option group or, if already present, the specified configuration is used to update the existing configuration.

$sel:optionsToRemove:ModifyOptionGroup', modifyOptionGroup_optionsToRemove - Options in this list are removed from the option group.

$sel:applyImmediately:ModifyOptionGroup', modifyOptionGroup_applyImmediately - A value that indicates whether to apply the change immediately or during the next maintenance window for each instance associated with the option group.

$sel:optionGroupName:ModifyOptionGroup', modifyOptionGroup_optionGroupName - The name of the option group to be modified.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, can't be removed from an option group, and that option group can't be removed from a DB instance once it is associated with a DB instance

Request Lenses

modifyOptionGroup_optionsToInclude :: Lens' ModifyOptionGroup (Maybe [OptionConfiguration])

Options in this list are added to the option group or, if already present, the specified configuration is used to update the existing configuration.

modifyOptionGroup_optionsToRemove :: Lens' ModifyOptionGroup (Maybe [Text])

Options in this list are removed from the option group.

modifyOptionGroup_applyImmediately :: Lens' ModifyOptionGroup (Maybe Bool)

A value that indicates whether to apply the change immediately or during the next maintenance window for each instance associated with the option group.

modifyOptionGroup_optionGroupName :: Lens' ModifyOptionGroup Text

The name of the option group to be modified.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, can't be removed from an option group, and that option group can't be removed from a DB instance once it is associated with a DB instance

Destructuring the Response

data ModifyOptionGroupResponse

See: newModifyOptionGroupResponse smart constructor.

Constructors

ModifyOptionGroupResponse'

Fields

optionGroup :: Maybe OptionGroup

httpStatus :: Int

The response's http status code.

newModifyOptionGroupResponse

Arguments

:: Int
$sel:httpStatus:ModifyOptionGroupResponse'

-> ModifyOptionGroupResponse

Create a value of ModifyOptionGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:optionGroup:ModifyOptionGroupResponse', modifyOptionGroupResponse_optionGroup - Undocumented member.

$sel:httpStatus:ModifyOptionGroupResponse', modifyOptionGroupResponse_httpStatus - The response's http status code.

Response Lenses

modifyOptionGroupResponse_optionGroup :: Lens' ModifyOptionGroupResponse (Maybe OptionGroup)

Undocumented member.

modifyOptionGroupResponse_httpStatus :: Lens' ModifyOptionGroupResponse Int

The response's http status code.