Network.AWS.RDS.PromoteReadReplica

Description

Promotes a read replica DB instance to a standalone DB instance.

Backup duration is a function of the amount of changes to the database since the previous backup. If you plan to promote a read replica to a standalone instance, we recommend that you enable backups and complete at least one backup prior to promotion. In addition, a read replica cannot be promoted to a standalone instance when it is in the backing-up status. If you have enabled backups on your read replica, configure the automated backup window so that daily backups do not interfere with read replica promotion.

This command doesn't apply to Aurora MySQL and Aurora PostgreSQL.

Creating a Request

data PromoteReadReplica

See: newPromoteReadReplica smart constructor.

Constructors

PromoteReadReplica'

Fields

preferredBackupWindow :: Maybe Text

The daily time range during which automated backups are created if automated backups are enabled, using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon RDS User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

backupRetentionPeriod :: Maybe Int

The number of days for which automated backups are retained. Setting this parameter to a positive number enables backups. Setting this parameter to 0 disables automated backups.

Default: 1

Constraints:

Must be a value from 0 to 35.

Can't be set to 0 if the DB instance is a source to read replicas.

dbInstanceIdentifier :: Text

The DB instance identifier. This value is stored as a lowercase string.

Constraints:

Must match the identifier of an existing read replica DB instance.

Example: mydbinstance

newPromoteReadReplica

Arguments

:: Text
$sel:dbInstanceIdentifier:PromoteReadReplica'

-> PromoteReadReplica

Create a value of PromoteReadReplica with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:preferredBackupWindow:PromoteReadReplica', promoteReadReplica_preferredBackupWindow - The daily time range during which automated backups are created if automated backups are enabled, using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon RDS User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

$sel:backupRetentionPeriod:PromoteReadReplica', promoteReadReplica_backupRetentionPeriod - The number of days for which automated backups are retained. Setting this parameter to a positive number enables backups. Setting this parameter to 0 disables automated backups.

Default: 1

Constraints:

Must be a value from 0 to 35.

Can't be set to 0 if the DB instance is a source to read replicas.

$sel:dbInstanceIdentifier:PromoteReadReplica', promoteReadReplica_dbInstanceIdentifier - The DB instance identifier. This value is stored as a lowercase string.

Constraints:

Must match the identifier of an existing read replica DB instance.

Example: mydbinstance

Request Lenses

promoteReadReplica_preferredBackupWindow :: Lens' PromoteReadReplica (Maybe Text)

The daily time range during which automated backups are created if automated backups are enabled, using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon RDS User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

promoteReadReplica_backupRetentionPeriod :: Lens' PromoteReadReplica (Maybe Int)

The number of days for which automated backups are retained. Setting this parameter to a positive number enables backups. Setting this parameter to 0 disables automated backups.

Default: 1

Constraints:

Must be a value from 0 to 35.

Can't be set to 0 if the DB instance is a source to read replicas.

promoteReadReplica_dbInstanceIdentifier :: Lens' PromoteReadReplica Text

The DB instance identifier. This value is stored as a lowercase string.

Constraints:

Must match the identifier of an existing read replica DB instance.

Example: mydbinstance

Destructuring the Response

data PromoteReadReplicaResponse

See: newPromoteReadReplicaResponse smart constructor.

Constructors

PromoteReadReplicaResponse'

Fields

dbInstance :: Maybe DBInstance

httpStatus :: Int

The response's http status code.

newPromoteReadReplicaResponse

Arguments

:: Int
$sel:httpStatus:PromoteReadReplicaResponse'

-> PromoteReadReplicaResponse

Create a value of PromoteReadReplicaResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbInstance:PromoteReadReplicaResponse', promoteReadReplicaResponse_dbInstance - Undocumented member.

$sel:httpStatus:PromoteReadReplicaResponse', promoteReadReplicaResponse_httpStatus - The response's http status code.

Response Lenses

promoteReadReplicaResponse_dbInstance :: Lens' PromoteReadReplicaResponse (Maybe DBInstance)

Undocumented member.

promoteReadReplicaResponse_httpStatus :: Lens' PromoteReadReplicaResponse Int

The response's http status code.