Network.AWS.RDS.ResetDBClusterParameterGroup

Description

Modifies the parameters of a DB cluster parameter group to the default value. To reset specific parameters submit a list of the following: ParameterName and ApplyMethod. To reset the entire DB cluster parameter group, specify the DBClusterParameterGroupName and ResetAllParameters parameters.

When resetting the entire group, dynamic parameters are updated immediately and static parameters are set to pending-reboot to take effect on the next DB instance restart or RebootDBInstance request. You must call RebootDBInstance for every DB instance in your DB cluster that you want the updated static parameter to apply to.

For more information on Amazon Aurora, see What Is Amazon Aurora? in the Amazon Aurora User Guide.

This action only applies to Aurora DB clusters.

Creating a Request

data ResetDBClusterParameterGroup

See: newResetDBClusterParameterGroup smart constructor.

Constructors

ResetDBClusterParameterGroup'

Fields

resetAllParameters :: Maybe Bool

A value that indicates whether to reset all parameters in the DB cluster parameter group to their default values. You can't use this parameter if there is a list of parameter names specified for the Parameters parameter.

parameters :: Maybe [Parameter]

A list of parameter names in the DB cluster parameter group to reset to the default values. You can't use this parameter if the ResetAllParameters parameter is enabled.

dbClusterParameterGroupName :: Text

The name of the DB cluster parameter group to reset.

newResetDBClusterParameterGroup

Arguments

:: Text
$sel:dbClusterParameterGroupName:ResetDBClusterParameterGroup'

-> ResetDBClusterParameterGroup

Create a value of ResetDBClusterParameterGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resetAllParameters:ResetDBClusterParameterGroup', resetDBClusterParameterGroup_resetAllParameters - A value that indicates whether to reset all parameters in the DB cluster parameter group to their default values. You can't use this parameter if there is a list of parameter names specified for the Parameters parameter.

$sel:parameters:ResetDBClusterParameterGroup', resetDBClusterParameterGroup_parameters - A list of parameter names in the DB cluster parameter group to reset to the default values. You can't use this parameter if the ResetAllParameters parameter is enabled.

$sel:dbClusterParameterGroupName:ResetDBClusterParameterGroup', resetDBClusterParameterGroup_dbClusterParameterGroupName - The name of the DB cluster parameter group to reset.

Request Lenses

resetDBClusterParameterGroup_resetAllParameters :: Lens' ResetDBClusterParameterGroup (Maybe Bool)

A value that indicates whether to reset all parameters in the DB cluster parameter group to their default values. You can't use this parameter if there is a list of parameter names specified for the Parameters parameter.

resetDBClusterParameterGroup_parameters :: Lens' ResetDBClusterParameterGroup (Maybe [Parameter])

A list of parameter names in the DB cluster parameter group to reset to the default values. You can't use this parameter if the ResetAllParameters parameter is enabled.

resetDBClusterParameterGroup_dbClusterParameterGroupName :: Lens' ResetDBClusterParameterGroup Text

The name of the DB cluster parameter group to reset.

Destructuring the Response

data DBClusterParameterGroupNameMessage

See: newDBClusterParameterGroupNameMessage smart constructor.

Constructors

DBClusterParameterGroupNameMessage'

Fields

dbClusterParameterGroupName :: Maybe Text

The name of the DB cluster parameter group.

Constraints:

Must be 1 to 255 letters or numbers.

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

This value is stored as a lowercase string.

newDBClusterParameterGroupNameMessage :: DBClusterParameterGroupNameMessage

Create a value of DBClusterParameterGroupNameMessage with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbClusterParameterGroupName:DBClusterParameterGroupNameMessage', dbClusterParameterGroupNameMessage_dbClusterParameterGroupName - The name of the DB cluster parameter group.

Constraints:

Must be 1 to 255 letters or numbers.

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

This value is stored as a lowercase string.

Response Lenses

dbClusterParameterGroupNameMessage_dbClusterParameterGroupName :: Lens' DBClusterParameterGroupNameMessage (Maybe Text)

The name of the DB cluster parameter group.

Constraints:

Must be 1 to 255 letters or numbers.

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

This value is stored as a lowercase string.