Network.AWS.RDS.RestoreDBClusterFromS

Description

Creates an Amazon Aurora DB cluster from MySQL data stored in an Amazon S3 bucket. Amazon RDS must be authorized to access the Amazon S3 bucket and the data must be created using the Percona XtraBackup utility as described in Migrating Data from MySQL by Using an Amazon S3 Bucket in the Amazon Aurora User Guide.

This action only restores the DB cluster, not the DB instances for that DB cluster. You must invoke the CreateDBInstance action to create DB instances for the restored DB cluster, specifying the identifier of the restored DB cluster in DBClusterIdentifier. You can create DB instances only after the RestoreDBClusterFromS3 action has completed and the DB cluster is available.

For more information on Amazon Aurora, see What Is Amazon Aurora? in the Amazon Aurora User Guide.

This action only applies to Aurora DB clusters. The source DB engine must be MySQL.

Creating a Request

data RestoreDBClusterFromS

See: newRestoreDBClusterFromS smart constructor.

Constructors

RestoreDBClusterFromS'

Fields

engineVersion :: Maybe Text

The version number of the database engine to use.

To list all of the available engine versions for aurora (for MySQL 5.6-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-mysql (for MySQL 5.7-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora-mysql --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-postgresql, use the following command:

aws rds describe-db-engine-versions --engine aurora-postgresql --query "DBEngineVersions[].EngineVersion"

Aurora MySQL

Example: 5.6.10a, 5.6.mysql_aurora.1.19.2, 5.7.12, 5.7.mysql_aurora.2.04.5

Aurora PostgreSQL

Example: 9.6.3, 10.7

deletionProtection :: Maybe Bool

A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

storageEncrypted :: Maybe Bool

A value that indicates whether the restored DB cluster is encrypted.

dbSubnetGroupName :: Maybe Text

A DB subnet group to associate with the restored DB cluster.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

domain :: Maybe Text

Specify the Active Directory directory ID to restore the DB cluster in. The domain must be created prior to this operation.

For Amazon Aurora DB clusters, Amazon RDS can use Kerberos Authentication to authenticate users that connect to the DB cluster. For more information, see Kerberos Authentication in the Amazon Aurora User Guide.

backtrackWindow :: Maybe Integer

The target backtrack window, in seconds. To disable backtracking, set this value to 0.

Currently, Backtrack is only supported for Aurora MySQL DB clusters.

Default: 0

Constraints:

If specified, this value must be set to a number from 0 to 259,200 (72 hours).

preferredMaintenanceWindow :: Maybe Text

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon Aurora User Guide.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

availabilityZones :: Maybe [Text]

A list of Availability Zones (AZs) where instances in the restored DB cluster can be created.

characterSetName :: Maybe Text

A value that indicates that the restored DB cluster should be associated with the specified CharacterSet.

kmsKeyId :: Maybe Text

The Amazon Web Services KMS key identifier for an encrypted DB cluster.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK). To use a CMK in a different Amazon Web Services account, specify the key ARN or alias ARN.

If the StorageEncrypted parameter is enabled, and you do not specify a value for the KmsKeyId parameter, then Amazon RDS will use your default CMK. There is a default CMK for your Amazon Web Services account. Your Amazon Web Services account has a different default CMK for each Amazon Web Services Region.

preferredBackupWindow :: Maybe Text

The daily time range during which automated backups are created if automated backups are enabled using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region. To view the time blocks available, see Backup window in the Amazon Aurora User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

backupRetentionPeriod :: Maybe Int

The number of days for which automated backups of the restored DB cluster are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35

vpcSecurityGroupIds :: Maybe [Text]

A list of EC2 VPC security groups to associate with the restored DB cluster.

databaseName :: Maybe Text

The database name for the restored DB cluster.

dbClusterParameterGroupName :: Maybe Text

The name of the DB cluster parameter group to associate with the restored DB cluster. If this argument is omitted, default.aurora5.6 is used.

Constraints:

If supplied, must match the name of an existing DBClusterParameterGroup.

s3Prefix :: Maybe Text

The prefix for all of the file names that contain the data used to create the Amazon Aurora DB cluster. If you do not specify a SourceS3Prefix value, then the Amazon Aurora DB cluster is created by using all of the files in the Amazon S3 bucket.

optionGroupName :: Maybe Text

A value that indicates that the restored DB cluster should be associated with the specified option group.

Permanent options can't be removed from an option group. An option group can't be removed from a DB cluster once it is associated with a DB cluster.

copyTagsToSnapshot :: Maybe Bool

A value that indicates whether to copy all tags from the restored DB cluster to snapshots of the restored DB cluster. The default is not to copy them.

domainIAMRoleName :: Maybe Text

Specify the name of the IAM role to be used when making API calls to the Directory Service.

tags :: Maybe [Tag]

port :: Maybe Int

The port number on which the instances in the restored DB cluster accept connections.

Default: 3306

enableIAMDatabaseAuthentication :: Maybe Bool

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information, see IAM Database Authentication in the Amazon Aurora User Guide.

enableCloudwatchLogsExports :: Maybe [Text]

The list of logs that the restored DB cluster is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon Aurora User Guide.

dbClusterIdentifier :: Text

The name of the DB cluster to create from the source data in the Amazon S3 bucket. This parameter isn't case-sensitive.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: my-cluster1

engine :: Text

The name of the database engine to be used for this DB cluster.

Valid Values: aurora (for MySQL 5.6-compatible Aurora), aurora-mysql (for MySQL 5.7-compatible Aurora), and aurora-postgresql

masterUsername :: Text

The name of the master user for the restored DB cluster.

Constraints:

Must be 1 to 16 letters or numbers.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

masterUserPassword :: Text

The password for the master database user. This password can contain any printable ASCII character except "/", """, or "@".

Constraints: Must contain from 8 to 41 characters.

sourceEngine :: Text

The identifier for the database engine that was backed up to create the files stored in the Amazon S3 bucket.

Valid values: mysql

sourceEngineVersion :: Text

The version of the database that the backup files were created from.

MySQL versions 5.5, 5.6, and 5.7 are supported.

Example: 5.6.40, 5.7.28

s3BucketName :: Text

The name of the Amazon S3 bucket that contains the data used to create the Amazon Aurora DB cluster.

s3IngestionRoleArn :: Text

The Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that authorizes Amazon RDS to access the Amazon S3 bucket on your behalf.

newRestoreDBClusterFromS

Arguments

:: Text
$sel:dbClusterIdentifier:RestoreDBClusterFromS'

-> Text
$sel:engine:RestoreDBClusterFromS'

-> Text
$sel:masterUsername:RestoreDBClusterFromS'

-> Text
$sel:masterUserPassword:RestoreDBClusterFromS'

-> Text
$sel:sourceEngine:RestoreDBClusterFromS'

-> Text
$sel:sourceEngineVersion:RestoreDBClusterFromS'

-> Text
$sel:s3BucketName:RestoreDBClusterFromS'

-> Text
$sel:s3IngestionRoleArn:RestoreDBClusterFromS'

-> RestoreDBClusterFromS

Create a value of RestoreDBClusterFromS with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:RestoreDBClusterFromS', restoreDBClusterFromS_engineVersion - The version number of the database engine to use.

To list all of the available engine versions for aurora (for MySQL 5.6-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-mysql (for MySQL 5.7-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora-mysql --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-postgresql, use the following command:

aws rds describe-db-engine-versions --engine aurora-postgresql --query "DBEngineVersions[].EngineVersion"

Aurora MySQL

Example: 5.6.10a, 5.6.mysql_aurora.1.19.2, 5.7.12, 5.7.mysql_aurora.2.04.5

Aurora PostgreSQL

Example: 9.6.3, 10.7

$sel:deletionProtection:RestoreDBClusterFromS', restoreDBClusterFromS_deletionProtection - A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

$sel:storageEncrypted:RestoreDBClusterFromS', restoreDBClusterFromS_storageEncrypted - A value that indicates whether the restored DB cluster is encrypted.

$sel:dbSubnetGroupName:RestoreDBClusterFromS', restoreDBClusterFromS_dbSubnetGroupName - A DB subnet group to associate with the restored DB cluster.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

$sel:domain:RestoreDBClusterFromS', restoreDBClusterFromS_domain - Specify the Active Directory directory ID to restore the DB cluster in. The domain must be created prior to this operation.

For Amazon Aurora DB clusters, Amazon RDS can use Kerberos Authentication to authenticate users that connect to the DB cluster. For more information, see Kerberos Authentication in the Amazon Aurora User Guide.

$sel:backtrackWindow:RestoreDBClusterFromS', restoreDBClusterFromS_backtrackWindow - The target backtrack window, in seconds. To disable backtracking, set this value to 0.

Currently, Backtrack is only supported for Aurora MySQL DB clusters.

Default: 0

Constraints:

If specified, this value must be set to a number from 0 to 259,200 (72 hours).

$sel:preferredMaintenanceWindow:RestoreDBClusterFromS', restoreDBClusterFromS_preferredMaintenanceWindow - The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon Aurora User Guide.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

$sel:availabilityZones:RestoreDBClusterFromS', restoreDBClusterFromS_availabilityZones - A list of Availability Zones (AZs) where instances in the restored DB cluster can be created.

$sel:characterSetName:RestoreDBClusterFromS', restoreDBClusterFromS_characterSetName - A value that indicates that the restored DB cluster should be associated with the specified CharacterSet.

$sel:kmsKeyId:RestoreDBClusterFromS', restoreDBClusterFromS_kmsKeyId - The Amazon Web Services KMS key identifier for an encrypted DB cluster.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK). To use a CMK in a different Amazon Web Services account, specify the key ARN or alias ARN.

If the StorageEncrypted parameter is enabled, and you do not specify a value for the KmsKeyId parameter, then Amazon RDS will use your default CMK. There is a default CMK for your Amazon Web Services account. Your Amazon Web Services account has a different default CMK for each Amazon Web Services Region.

$sel:preferredBackupWindow:RestoreDBClusterFromS', restoreDBClusterFromS_preferredBackupWindow - The daily time range during which automated backups are created if automated backups are enabled using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region. To view the time blocks available, see Backup window in the Amazon Aurora User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

$sel:backupRetentionPeriod:RestoreDBClusterFromS', restoreDBClusterFromS_backupRetentionPeriod - The number of days for which automated backups of the restored DB cluster are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35

$sel:vpcSecurityGroupIds:RestoreDBClusterFromS', restoreDBClusterFromS_vpcSecurityGroupIds - A list of EC2 VPC security groups to associate with the restored DB cluster.

$sel:databaseName:RestoreDBClusterFromS', restoreDBClusterFromS_databaseName - The database name for the restored DB cluster.

$sel:dbClusterParameterGroupName:RestoreDBClusterFromS', restoreDBClusterFromS_dbClusterParameterGroupName - The name of the DB cluster parameter group to associate with the restored DB cluster. If this argument is omitted, default.aurora5.6 is used.

Constraints:

If supplied, must match the name of an existing DBClusterParameterGroup.

$sel:s3Prefix:RestoreDBClusterFromS', restoreDBClusterFromS_s3Prefix - The prefix for all of the file names that contain the data used to create the Amazon Aurora DB cluster. If you do not specify a SourceS3Prefix value, then the Amazon Aurora DB cluster is created by using all of the files in the Amazon S3 bucket.

$sel:optionGroupName:RestoreDBClusterFromS', restoreDBClusterFromS_optionGroupName - A value that indicates that the restored DB cluster should be associated with the specified option group.

Permanent options can't be removed from an option group. An option group can't be removed from a DB cluster once it is associated with a DB cluster.

$sel:copyTagsToSnapshot:RestoreDBClusterFromS', restoreDBClusterFromS_copyTagsToSnapshot - A value that indicates whether to copy all tags from the restored DB cluster to snapshots of the restored DB cluster. The default is not to copy them.

$sel:domainIAMRoleName:RestoreDBClusterFromS', restoreDBClusterFromS_domainIAMRoleName - Specify the name of the IAM role to be used when making API calls to the Directory Service.

$sel:tags:RestoreDBClusterFromS', restoreDBClusterFromS_tags - Undocumented member.

$sel:port:RestoreDBClusterFromS', restoreDBClusterFromS_port - The port number on which the instances in the restored DB cluster accept connections.

Default: 3306

$sel:enableIAMDatabaseAuthentication:RestoreDBClusterFromS', restoreDBClusterFromS_enableIAMDatabaseAuthentication - A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information, see IAM Database Authentication in the Amazon Aurora User Guide.

$sel:enableCloudwatchLogsExports:RestoreDBClusterFromS', restoreDBClusterFromS_enableCloudwatchLogsExports - The list of logs that the restored DB cluster is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon Aurora User Guide.

$sel:dbClusterIdentifier:RestoreDBClusterFromS', restoreDBClusterFromS_dbClusterIdentifier - The name of the DB cluster to create from the source data in the Amazon S3 bucket. This parameter isn't case-sensitive.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: my-cluster1

$sel:engine:RestoreDBClusterFromS', restoreDBClusterFromS_engine - The name of the database engine to be used for this DB cluster.

Valid Values: aurora (for MySQL 5.6-compatible Aurora), aurora-mysql (for MySQL 5.7-compatible Aurora), and aurora-postgresql

$sel:masterUsername:RestoreDBClusterFromS', restoreDBClusterFromS_masterUsername - The name of the master user for the restored DB cluster.

Constraints:

Must be 1 to 16 letters or numbers.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

$sel:masterUserPassword:RestoreDBClusterFromS', restoreDBClusterFromS_masterUserPassword - The password for the master database user. This password can contain any printable ASCII character except "/", """, or "@".

Constraints: Must contain from 8 to 41 characters.

$sel:sourceEngine:RestoreDBClusterFromS', restoreDBClusterFromS_sourceEngine - The identifier for the database engine that was backed up to create the files stored in the Amazon S3 bucket.

Valid values: mysql

$sel:sourceEngineVersion:RestoreDBClusterFromS', restoreDBClusterFromS_sourceEngineVersion - The version of the database that the backup files were created from.

MySQL versions 5.5, 5.6, and 5.7 are supported.

Example: 5.6.40, 5.7.28

$sel:s3BucketName:RestoreDBClusterFromS', restoreDBClusterFromS_s3BucketName - The name of the Amazon S3 bucket that contains the data used to create the Amazon Aurora DB cluster.

$sel:s3IngestionRoleArn:RestoreDBClusterFromS', restoreDBClusterFromS_s3IngestionRoleArn - The Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that authorizes Amazon RDS to access the Amazon S3 bucket on your behalf.

Request Lenses

restoreDBClusterFromS_engineVersion :: Lens' RestoreDBClusterFromS (Maybe Text)

The version number of the database engine to use.

To list all of the available engine versions for aurora (for MySQL 5.6-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-mysql (for MySQL 5.7-compatible Aurora), use the following command:

aws rds describe-db-engine-versions --engine aurora-mysql --query "DBEngineVersions[].EngineVersion"

To list all of the available engine versions for aurora-postgresql, use the following command:

aws rds describe-db-engine-versions --engine aurora-postgresql --query "DBEngineVersions[].EngineVersion"

Aurora MySQL

Example: 5.6.10a, 5.6.mysql_aurora.1.19.2, 5.7.12, 5.7.mysql_aurora.2.04.5

Aurora PostgreSQL

Example: 9.6.3, 10.7

restoreDBClusterFromS_deletionProtection :: Lens' RestoreDBClusterFromS (Maybe Bool)

A value that indicates whether the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled.

restoreDBClusterFromS_storageEncrypted :: Lens' RestoreDBClusterFromS (Maybe Bool)

A value that indicates whether the restored DB cluster is encrypted.

restoreDBClusterFromS_dbSubnetGroupName :: Lens' RestoreDBClusterFromS (Maybe Text)

A DB subnet group to associate with the restored DB cluster.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

restoreDBClusterFromS_domain :: Lens' RestoreDBClusterFromS (Maybe Text)

Specify the Active Directory directory ID to restore the DB cluster in. The domain must be created prior to this operation.

For Amazon Aurora DB clusters, Amazon RDS can use Kerberos Authentication to authenticate users that connect to the DB cluster. For more information, see Kerberos Authentication in the Amazon Aurora User Guide.

restoreDBClusterFromS_backtrackWindow :: Lens' RestoreDBClusterFromS (Maybe Integer)

The target backtrack window, in seconds. To disable backtracking, set this value to 0.

Currently, Backtrack is only supported for Aurora MySQL DB clusters.

Default: 0

Constraints:

If specified, this value must be set to a number from 0 to 259,200 (72 hours).

restoreDBClusterFromS_preferredMaintenanceWindow :: Lens' RestoreDBClusterFromS (Maybe Text)

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

Format: ddd:hh24:mi-ddd:hh24:mi

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region, occurring on a random day of the week. To see the time blocks available, see Adjusting the Preferred Maintenance Window in the Amazon Aurora User Guide.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Constraints: Minimum 30-minute window.

restoreDBClusterFromS_availabilityZones :: Lens' RestoreDBClusterFromS (Maybe [Text])

A list of Availability Zones (AZs) where instances in the restored DB cluster can be created.

restoreDBClusterFromS_characterSetName :: Lens' RestoreDBClusterFromS (Maybe Text)

A value that indicates that the restored DB cluster should be associated with the specified CharacterSet.

restoreDBClusterFromS_kmsKeyId :: Lens' RestoreDBClusterFromS (Maybe Text)

The Amazon Web Services KMS key identifier for an encrypted DB cluster.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK). To use a CMK in a different Amazon Web Services account, specify the key ARN or alias ARN.

If the StorageEncrypted parameter is enabled, and you do not specify a value for the KmsKeyId parameter, then Amazon RDS will use your default CMK. There is a default CMK for your Amazon Web Services account. Your Amazon Web Services account has a different default CMK for each Amazon Web Services Region.

restoreDBClusterFromS_preferredBackupWindow :: Lens' RestoreDBClusterFromS (Maybe Text)

The daily time range during which automated backups are created if automated backups are enabled using the BackupRetentionPeriod parameter.

The default is a 30-minute window selected at random from an 8-hour block of time for each Amazon Web Services Region. To view the time blocks available, see Backup window in the Amazon Aurora User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

restoreDBClusterFromS_backupRetentionPeriod :: Lens' RestoreDBClusterFromS (Maybe Int)

The number of days for which automated backups of the restored DB cluster are retained. You must specify a minimum value of 1.

Default: 1

Constraints:

Must be a value from 1 to 35

restoreDBClusterFromS_vpcSecurityGroupIds :: Lens' RestoreDBClusterFromS (Maybe [Text])

A list of EC2 VPC security groups to associate with the restored DB cluster.

restoreDBClusterFromS_databaseName :: Lens' RestoreDBClusterFromS (Maybe Text)

The database name for the restored DB cluster.

restoreDBClusterFromS_dbClusterParameterGroupName :: Lens' RestoreDBClusterFromS (Maybe Text)

The name of the DB cluster parameter group to associate with the restored DB cluster. If this argument is omitted, default.aurora5.6 is used.

Constraints:

If supplied, must match the name of an existing DBClusterParameterGroup.

restoreDBClusterFromS_s3Prefix :: Lens' RestoreDBClusterFromS (Maybe Text)

The prefix for all of the file names that contain the data used to create the Amazon Aurora DB cluster. If you do not specify a SourceS3Prefix value, then the Amazon Aurora DB cluster is created by using all of the files in the Amazon S3 bucket.

restoreDBClusterFromS_optionGroupName :: Lens' RestoreDBClusterFromS (Maybe Text)

A value that indicates that the restored DB cluster should be associated with the specified option group.

Permanent options can't be removed from an option group. An option group can't be removed from a DB cluster once it is associated with a DB cluster.

restoreDBClusterFromS_copyTagsToSnapshot :: Lens' RestoreDBClusterFromS (Maybe Bool)

A value that indicates whether to copy all tags from the restored DB cluster to snapshots of the restored DB cluster. The default is not to copy them.

restoreDBClusterFromS_domainIAMRoleName :: Lens' RestoreDBClusterFromS (Maybe Text)

Specify the name of the IAM role to be used when making API calls to the Directory Service.

restoreDBClusterFromS_tags :: Lens' RestoreDBClusterFromS (Maybe [Tag])

Undocumented member.

restoreDBClusterFromS_port :: Lens' RestoreDBClusterFromS (Maybe Int)

The port number on which the instances in the restored DB cluster accept connections.

Default: 3306

restoreDBClusterFromS_enableIAMDatabaseAuthentication :: Lens' RestoreDBClusterFromS (Maybe Bool)

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information, see IAM Database Authentication in the Amazon Aurora User Guide.

restoreDBClusterFromS_enableCloudwatchLogsExports :: Lens' RestoreDBClusterFromS (Maybe [Text])

The list of logs that the restored DB cluster is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon Aurora User Guide.

restoreDBClusterFromS_dbClusterIdentifier :: Lens' RestoreDBClusterFromS Text

The name of the DB cluster to create from the source data in the Amazon S3 bucket. This parameter isn't case-sensitive.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: my-cluster1

restoreDBClusterFromS_engine :: Lens' RestoreDBClusterFromS Text

The name of the database engine to be used for this DB cluster.

Valid Values: aurora (for MySQL 5.6-compatible Aurora), aurora-mysql (for MySQL 5.7-compatible Aurora), and aurora-postgresql

restoreDBClusterFromS_masterUsername :: Lens' RestoreDBClusterFromS Text

The name of the master user for the restored DB cluster.

Constraints:

Must be 1 to 16 letters or numbers.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

restoreDBClusterFromS_masterUserPassword :: Lens' RestoreDBClusterFromS Text

The password for the master database user. This password can contain any printable ASCII character except "/", """, or "@".

Constraints: Must contain from 8 to 41 characters.

restoreDBClusterFromS_sourceEngine :: Lens' RestoreDBClusterFromS Text

The identifier for the database engine that was backed up to create the files stored in the Amazon S3 bucket.

Valid values: mysql

restoreDBClusterFromS_sourceEngineVersion :: Lens' RestoreDBClusterFromS Text

The version of the database that the backup files were created from.

MySQL versions 5.5, 5.6, and 5.7 are supported.

Example: 5.6.40, 5.7.28

restoreDBClusterFromS_s3BucketName :: Lens' RestoreDBClusterFromS Text

The name of the Amazon S3 bucket that contains the data used to create the Amazon Aurora DB cluster.

restoreDBClusterFromS_s3IngestionRoleArn :: Lens' RestoreDBClusterFromS Text

The Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) role that authorizes Amazon RDS to access the Amazon S3 bucket on your behalf.

Destructuring the Response

data RestoreDBClusterFromSResponse

See: newRestoreDBClusterFromSResponse smart constructor.

Constructors

RestoreDBClusterFromSResponse'

Fields

dbCluster :: Maybe DBCluster

httpStatus :: Int

The response's http status code.

newRestoreDBClusterFromSResponse

Arguments

:: Int
$sel:httpStatus:RestoreDBClusterFromSResponse'

-> RestoreDBClusterFromSResponse

Create a value of RestoreDBClusterFromSResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbCluster:RestoreDBClusterFromSResponse', restoreDBClusterFromSResponse_dbCluster - Undocumented member.

$sel:httpStatus:RestoreDBClusterFromSResponse', restoreDBClusterFromSResponse_httpStatus - The response's http status code.

Response Lenses

restoreDBClusterFromSResponse_dbCluster :: Lens' RestoreDBClusterFromSResponse (Maybe DBCluster)

Undocumented member.

restoreDBClusterFromSResponse_httpStatus :: Lens' RestoreDBClusterFromSResponse Int

The response's http status code.