Network.AWS.RDS.RestoreDBInstanceFromDBSnapshot

Description

Creates a new DB instance from a DB snapshot. The target database is created from the source database restore point with most of the source's original configuration, including the default security group and DB parameter group. By default, the new DB instance is created as a Single-AZ deployment, except when the instance is a SQL Server instance that has an option group associated with mirroring. In this case, the instance becomes a Multi-AZ deployment, not a Single-AZ deployment.

If you want to replace your original DB instance with the new, restored DB instance, then rename your original DB instance before you call the RestoreDBInstanceFromDBSnapshot action. RDS doesn't allow two DB instances with the same name. After you have renamed your original DB instance with a different identifier, then you can pass the original name of the DB instance as the DBInstanceIdentifier in the call to the RestoreDBInstanceFromDBSnapshot action. The result is that you replace the original DB instance with the DB instance created from the snapshot.

If you are restoring from a shared manual DB snapshot, the DBSnapshotIdentifier must be the ARN of the shared DB snapshot.

This command doesn't apply to Aurora MySQL and Aurora PostgreSQL. For Aurora, use RestoreDBClusterFromSnapshot.

Creating a Request

data RestoreDBInstanceFromDBSnapshot

See: newRestoreDBInstanceFromDBSnapshot smart constructor.

Constructors

RestoreDBInstanceFromDBSnapshot'

Fields

deletionProtection :: Maybe Bool

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

publiclyAccessible :: Maybe Bool

A value that indicates whether the DB instance is publicly accessible.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

For more information, see CreateDBInstance.

autoMinorVersionUpgrade :: Maybe Bool

A value that indicates whether minor version upgrades are applied automatically to the DB instance during the maintenance window.

dbSubnetGroupName :: Maybe Text

The DB subnet group name to use for the new instance.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

iops :: Maybe Int

Specifies the amount of provisioned IOPS for the DB instance, expressed in I/O operations per second. If this parameter isn't specified, the IOPS value is taken from the backup. If this parameter is set to 0, the new instance is converted to a non-PIOPS instance. The conversion takes additional time, though your DB instance is available for connections before the conversion starts.

The provisioned IOPS value must follow the requirements for your database engine. For more information, see Amazon RDS Provisioned IOPS Storage to Improve Performance in the Amazon RDS User Guide.

Constraints: Must be an integer greater than 1000.

domain :: Maybe Text

Specify the Active Directory directory ID to restore the DB instance in. The domain must be created prior to this operation. Currently, only MySQL, Microsoft SQL Server, Oracle, and PostgreSQL DB instances can be created in an Active Directory Domain.

For more information, see Kerberos Authentication in the Amazon RDS User Guide.

enableCustomerOwnedIp :: Maybe Bool

A value that indicates whether to enable a customer-owned IP address (CoIP) for an RDS on Outposts DB instance.

A CoIP provides local or external connectivity to resources in your Outpost subnets through your on-premises network. For some use cases, a CoIP can provide lower latency for connections to the DB instance from outside of its virtual private cloud (VPC) on your local network.

For more information about RDS on Outposts, see Working with Amazon RDS on Amazon Web Services Outposts in the Amazon RDS User Guide.

For more information about CoIPs, see Customer-owned IP addresses in the Amazon Web Services Outposts User Guide.

engine :: Maybe Text

The database engine to use for the new instance.

Default: The same as source

Constraint: Must be compatible with the engine of the source. For example, you can restore a MariaDB 10.1 DB instance from a MySQL 5.6 snapshot.

Valid Values:

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

tdeCredentialPassword :: Maybe Text

The password for the given ARN from the key store in order to access the device.

processorFeatures :: Maybe [ProcessorFeature]

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

dbInstanceClass :: Maybe Text

The compute and memory capacity of the Amazon RDS DB instance, for example, db.m4.large. Not all DB instance classes are available in all Amazon Web Services Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

Default: The same DBInstanceClass as the original DB instance.

licenseModel :: Maybe Text

License model information for the restored DB instance.

Default: Same as source.

Valid values: license-included | bring-your-own-license | general-public-license

dbParameterGroupName :: Maybe Text

The name of the DB parameter group to associate with this DB instance.

If you do not specify a value for DBParameterGroupName, then the default DBParameterGroup for the specified DB engine is used.

Constraints:

If supplied, must match the name of an existing DBParameterGroup.

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

availabilityZone :: Maybe Text

The Availability Zone (AZ) where the DB instance will be created.

Default: A random, system-chosen Availability Zone.

Constraint: You can't specify the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

Example: us-east-1a

vpcSecurityGroupIds :: Maybe [Text]

A list of EC2 VPC security groups to associate with this DB instance.

Default: The default EC2 VPC security group for the DB subnet group's VPC.

multiAZ :: Maybe Bool

A value that indicates whether the DB instance is a Multi-AZ deployment.

Constraint: You can't specify the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

optionGroupName :: Maybe Text

The name of the option group to be used for the restored DB instance.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, can't be removed from an option group, and that option group can't be removed from a DB instance once it is associated with a DB instance

copyTagsToSnapshot :: Maybe Bool

A value that indicates whether to copy all tags from the restored DB instance to snapshots of the DB instance. By default, tags are not copied.

tdeCredentialArn :: Maybe Text

The ARN from the key store with which to associate the instance for TDE encryption.

domainIAMRoleName :: Maybe Text

Specify the name of the IAM role to be used when making API calls to the Directory Service.

tags :: Maybe [Tag]

port :: Maybe Int

The port number on which the database accepts connections.

Default: The same port as the original DB instance

Constraints: Value must be 1150-65535

enableIAMDatabaseAuthentication :: Maybe Bool

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information about IAM database authentication, see IAM Database Authentication for MySQL and PostgreSQL in the Amazon RDS User Guide.

useDefaultProcessorFeatures :: Maybe Bool

A value that indicates whether the DB instance class of the DB instance uses its default processor features.

storageType :: Maybe Text

Specifies the storage type to be associated with the DB instance.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified, otherwise gp2

enableCloudwatchLogsExports :: Maybe [Text]

The list of logs that the restored DB instance is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon RDS User Guide.

dbName :: Maybe Text

The database name for the restored DB instance.

This parameter doesn't apply to the MySQL, PostgreSQL, or MariaDB engines.

dbInstanceIdentifier :: Text

Name of the DB instance to create from the DB snapshot. This parameter isn't case-sensitive.

Constraints:

Must contain from 1 to 63 numbers, letters, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: my-snapshot-id

dbSnapshotIdentifier :: Text

The identifier for the DB snapshot to restore from.

Constraints:

Must match the identifier of an existing DBSnapshot.

If you are restoring from a shared manual DB snapshot, the DBSnapshotIdentifier must be the ARN of the shared DB snapshot.

newRestoreDBInstanceFromDBSnapshot

Arguments

:: Text
$sel:dbInstanceIdentifier:RestoreDBInstanceFromDBSnapshot'

-> Text
$sel:dbSnapshotIdentifier:RestoreDBInstanceFromDBSnapshot'

-> RestoreDBInstanceFromDBSnapshot

Create a value of RestoreDBInstanceFromDBSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deletionProtection:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_deletionProtection - A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

$sel:publiclyAccessible:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_publiclyAccessible - A value that indicates whether the DB instance is publicly accessible.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

For more information, see CreateDBInstance.

$sel:autoMinorVersionUpgrade:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_autoMinorVersionUpgrade - A value that indicates whether minor version upgrades are applied automatically to the DB instance during the maintenance window.

$sel:dbSubnetGroupName:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_dbSubnetGroupName - The DB subnet group name to use for the new instance.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

$sel:iops:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_iops - Specifies the amount of provisioned IOPS for the DB instance, expressed in I/O operations per second. If this parameter isn't specified, the IOPS value is taken from the backup. If this parameter is set to 0, the new instance is converted to a non-PIOPS instance. The conversion takes additional time, though your DB instance is available for connections before the conversion starts.

The provisioned IOPS value must follow the requirements for your database engine. For more information, see Amazon RDS Provisioned IOPS Storage to Improve Performance in the Amazon RDS User Guide.

Constraints: Must be an integer greater than 1000.

$sel:domain:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_domain - Specify the Active Directory directory ID to restore the DB instance in. The domain must be created prior to this operation. Currently, only MySQL, Microsoft SQL Server, Oracle, and PostgreSQL DB instances can be created in an Active Directory Domain.

For more information, see Kerberos Authentication in the Amazon RDS User Guide.

$sel:enableCustomerOwnedIp:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_enableCustomerOwnedIp - A value that indicates whether to enable a customer-owned IP address (CoIP) for an RDS on Outposts DB instance.

A CoIP provides local or external connectivity to resources in your Outpost subnets through your on-premises network. For some use cases, a CoIP can provide lower latency for connections to the DB instance from outside of its virtual private cloud (VPC) on your local network.

For more information about RDS on Outposts, see Working with Amazon RDS on Amazon Web Services Outposts in the Amazon RDS User Guide.

For more information about CoIPs, see Customer-owned IP addresses in the Amazon Web Services Outposts User Guide.

$sel:engine:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_engine - The database engine to use for the new instance.

Default: The same as source

Constraint: Must be compatible with the engine of the source. For example, you can restore a MariaDB 10.1 DB instance from a MySQL 5.6 snapshot.

Valid Values:

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

$sel:tdeCredentialPassword:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_tdeCredentialPassword - The password for the given ARN from the key store in order to access the device.

$sel:processorFeatures:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_processorFeatures - The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

$sel:dbInstanceClass:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_dbInstanceClass - The compute and memory capacity of the Amazon RDS DB instance, for example, db.m4.large. Not all DB instance classes are available in all Amazon Web Services Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

Default: The same DBInstanceClass as the original DB instance.

$sel:licenseModel:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_licenseModel - License model information for the restored DB instance.

Default: Same as source.

Valid values: license-included | bring-your-own-license | general-public-license

$sel:dbParameterGroupName:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_dbParameterGroupName - The name of the DB parameter group to associate with this DB instance.

If you do not specify a value for DBParameterGroupName, then the default DBParameterGroup for the specified DB engine is used.

Constraints:

If supplied, must match the name of an existing DBParameterGroup.

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

$sel:availabilityZone:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_availabilityZone - The Availability Zone (AZ) where the DB instance will be created.

Default: A random, system-chosen Availability Zone.

Constraint: You can't specify the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

Example: us-east-1a

$sel:vpcSecurityGroupIds:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_vpcSecurityGroupIds - A list of EC2 VPC security groups to associate with this DB instance.

Default: The default EC2 VPC security group for the DB subnet group's VPC.

$sel:multiAZ:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_multiAZ - A value that indicates whether the DB instance is a Multi-AZ deployment.

Constraint: You can't specify the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

$sel:optionGroupName:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_optionGroupName - The name of the option group to be used for the restored DB instance.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, can't be removed from an option group, and that option group can't be removed from a DB instance once it is associated with a DB instance

$sel:copyTagsToSnapshot:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_copyTagsToSnapshot - A value that indicates whether to copy all tags from the restored DB instance to snapshots of the DB instance. By default, tags are not copied.

$sel:tdeCredentialArn:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_tdeCredentialArn - The ARN from the key store with which to associate the instance for TDE encryption.

$sel:domainIAMRoleName:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_domainIAMRoleName - Specify the name of the IAM role to be used when making API calls to the Directory Service.

$sel:tags:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_tags - Undocumented member.

$sel:port:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_port - The port number on which the database accepts connections.

Default: The same port as the original DB instance

Constraints: Value must be 1150-65535

$sel:enableIAMDatabaseAuthentication:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_enableIAMDatabaseAuthentication - A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information about IAM database authentication, see IAM Database Authentication for MySQL and PostgreSQL in the Amazon RDS User Guide.

$sel:useDefaultProcessorFeatures:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_useDefaultProcessorFeatures - A value that indicates whether the DB instance class of the DB instance uses its default processor features.

$sel:storageType:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_storageType - Specifies the storage type to be associated with the DB instance.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified, otherwise gp2

$sel:enableCloudwatchLogsExports:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_enableCloudwatchLogsExports - The list of logs that the restored DB instance is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon RDS User Guide.

$sel:dbName:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_dbName - The database name for the restored DB instance.

This parameter doesn't apply to the MySQL, PostgreSQL, or MariaDB engines.

$sel:dbInstanceIdentifier:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_dbInstanceIdentifier - Name of the DB instance to create from the DB snapshot. This parameter isn't case-sensitive.

Constraints:

Must contain from 1 to 63 numbers, letters, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: my-snapshot-id

$sel:dbSnapshotIdentifier:RestoreDBInstanceFromDBSnapshot', restoreDBInstanceFromDBSnapshot_dbSnapshotIdentifier - The identifier for the DB snapshot to restore from.

Constraints:

Must match the identifier of an existing DBSnapshot.

If you are restoring from a shared manual DB snapshot, the DBSnapshotIdentifier must be the ARN of the shared DB snapshot.

Request Lenses

restoreDBInstanceFromDBSnapshot_deletionProtection :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Bool)

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

restoreDBInstanceFromDBSnapshot_publiclyAccessible :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Bool)

A value that indicates whether the DB instance is publicly accessible.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

For more information, see CreateDBInstance.

restoreDBInstanceFromDBSnapshot_autoMinorVersionUpgrade :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Bool)

A value that indicates whether minor version upgrades are applied automatically to the DB instance during the maintenance window.

restoreDBInstanceFromDBSnapshot_dbSubnetGroupName :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The DB subnet group name to use for the new instance.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

restoreDBInstanceFromDBSnapshot_iops :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Int)

Specifies the amount of provisioned IOPS for the DB instance, expressed in I/O operations per second. If this parameter isn't specified, the IOPS value is taken from the backup. If this parameter is set to 0, the new instance is converted to a non-PIOPS instance. The conversion takes additional time, though your DB instance is available for connections before the conversion starts.

The provisioned IOPS value must follow the requirements for your database engine. For more information, see Amazon RDS Provisioned IOPS Storage to Improve Performance in the Amazon RDS User Guide.

Constraints: Must be an integer greater than 1000.

restoreDBInstanceFromDBSnapshot_domain :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

Specify the Active Directory directory ID to restore the DB instance in. The domain must be created prior to this operation. Currently, only MySQL, Microsoft SQL Server, Oracle, and PostgreSQL DB instances can be created in an Active Directory Domain.

For more information, see Kerberos Authentication in the Amazon RDS User Guide.

restoreDBInstanceFromDBSnapshot_enableCustomerOwnedIp :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Bool)

A value that indicates whether to enable a customer-owned IP address (CoIP) for an RDS on Outposts DB instance.

A CoIP provides local or external connectivity to resources in your Outpost subnets through your on-premises network. For some use cases, a CoIP can provide lower latency for connections to the DB instance from outside of its virtual private cloud (VPC) on your local network.

For more information about RDS on Outposts, see Working with Amazon RDS on Amazon Web Services Outposts in the Amazon RDS User Guide.

For more information about CoIPs, see Customer-owned IP addresses in the Amazon Web Services Outposts User Guide.

restoreDBInstanceFromDBSnapshot_engine :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The database engine to use for the new instance.

Default: The same as source

Constraint: Must be compatible with the engine of the source. For example, you can restore a MariaDB 10.1 DB instance from a MySQL 5.6 snapshot.

Valid Values:

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

restoreDBInstanceFromDBSnapshot_tdeCredentialPassword :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The password for the given ARN from the key store in order to access the device.

restoreDBInstanceFromDBSnapshot_processorFeatures :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe [ProcessorFeature])

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

restoreDBInstanceFromDBSnapshot_dbInstanceClass :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The compute and memory capacity of the Amazon RDS DB instance, for example, db.m4.large. Not all DB instance classes are available in all Amazon Web Services Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

Default: The same DBInstanceClass as the original DB instance.

restoreDBInstanceFromDBSnapshot_licenseModel :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

License model information for the restored DB instance.

Default: Same as source.

Valid values: license-included | bring-your-own-license | general-public-license

restoreDBInstanceFromDBSnapshot_dbParameterGroupName :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The name of the DB parameter group to associate with this DB instance.

If you do not specify a value for DBParameterGroupName, then the default DBParameterGroup for the specified DB engine is used.

Constraints:

If supplied, must match the name of an existing DBParameterGroup.

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

restoreDBInstanceFromDBSnapshot_availabilityZone :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The Availability Zone (AZ) where the DB instance will be created.

Default: A random, system-chosen Availability Zone.

Constraint: You can't specify the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

Example: us-east-1a

restoreDBInstanceFromDBSnapshot_vpcSecurityGroupIds :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe [Text])

A list of EC2 VPC security groups to associate with this DB instance.

Default: The default EC2 VPC security group for the DB subnet group's VPC.

restoreDBInstanceFromDBSnapshot_multiAZ :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Bool)

A value that indicates whether the DB instance is a Multi-AZ deployment.

Constraint: You can't specify the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

restoreDBInstanceFromDBSnapshot_optionGroupName :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The name of the option group to be used for the restored DB instance.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, can't be removed from an option group, and that option group can't be removed from a DB instance once it is associated with a DB instance

restoreDBInstanceFromDBSnapshot_copyTagsToSnapshot :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Bool)

A value that indicates whether to copy all tags from the restored DB instance to snapshots of the DB instance. By default, tags are not copied.

restoreDBInstanceFromDBSnapshot_tdeCredentialArn :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The ARN from the key store with which to associate the instance for TDE encryption.

restoreDBInstanceFromDBSnapshot_domainIAMRoleName :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

Specify the name of the IAM role to be used when making API calls to the Directory Service.

restoreDBInstanceFromDBSnapshot_tags :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe [Tag])

Undocumented member.

restoreDBInstanceFromDBSnapshot_port :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Int)

The port number on which the database accepts connections.

Default: The same port as the original DB instance

Constraints: Value must be 1150-65535

restoreDBInstanceFromDBSnapshot_enableIAMDatabaseAuthentication :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Bool)

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information about IAM database authentication, see IAM Database Authentication for MySQL and PostgreSQL in the Amazon RDS User Guide.

restoreDBInstanceFromDBSnapshot_useDefaultProcessorFeatures :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Bool)

A value that indicates whether the DB instance class of the DB instance uses its default processor features.

restoreDBInstanceFromDBSnapshot_storageType :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

Specifies the storage type to be associated with the DB instance.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified, otherwise gp2

restoreDBInstanceFromDBSnapshot_enableCloudwatchLogsExports :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe [Text])

The list of logs that the restored DB instance is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon RDS User Guide.

restoreDBInstanceFromDBSnapshot_dbName :: Lens' RestoreDBInstanceFromDBSnapshot (Maybe Text)

The database name for the restored DB instance.

This parameter doesn't apply to the MySQL, PostgreSQL, or MariaDB engines.

restoreDBInstanceFromDBSnapshot_dbInstanceIdentifier :: Lens' RestoreDBInstanceFromDBSnapshot Text

Name of the DB instance to create from the DB snapshot. This parameter isn't case-sensitive.

Constraints:

Must contain from 1 to 63 numbers, letters, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Example: my-snapshot-id

restoreDBInstanceFromDBSnapshot_dbSnapshotIdentifier :: Lens' RestoreDBInstanceFromDBSnapshot Text

The identifier for the DB snapshot to restore from.

Constraints:

Must match the identifier of an existing DBSnapshot.

If you are restoring from a shared manual DB snapshot, the DBSnapshotIdentifier must be the ARN of the shared DB snapshot.

Destructuring the Response

data RestoreDBInstanceFromDBSnapshotResponse

See: newRestoreDBInstanceFromDBSnapshotResponse smart constructor.

Constructors

RestoreDBInstanceFromDBSnapshotResponse'

Fields

dbInstance :: Maybe DBInstance

httpStatus :: Int

The response's http status code.

newRestoreDBInstanceFromDBSnapshotResponse

Arguments

:: Int
$sel:httpStatus:RestoreDBInstanceFromDBSnapshotResponse'

-> RestoreDBInstanceFromDBSnapshotResponse

Create a value of RestoreDBInstanceFromDBSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbInstance:RestoreDBInstanceFromDBSnapshotResponse', restoreDBInstanceFromDBSnapshotResponse_dbInstance - Undocumented member.

$sel:httpStatus:RestoreDBInstanceFromDBSnapshotResponse', restoreDBInstanceFromDBSnapshotResponse_httpStatus - The response's http status code.

Response Lenses

restoreDBInstanceFromDBSnapshotResponse_dbInstance :: Lens' RestoreDBInstanceFromDBSnapshotResponse (Maybe DBInstance)

Undocumented member.

restoreDBInstanceFromDBSnapshotResponse_httpStatus :: Lens' RestoreDBInstanceFromDBSnapshotResponse Int

The response's http status code.