Network.AWS.RDS.RestoreDBInstanceFromS3

Description

Amazon Relational Database Service (Amazon RDS) supports importing MySQL databases by using backup files. You can create a backup of your on-premises database, store it on Amazon Simple Storage Service (Amazon S3), and then restore the backup file onto a new Amazon RDS DB instance running MySQL. For more information, see Importing Data into an Amazon RDS MySQL DB Instance in the Amazon RDS User Guide.

Creating a Request

data RestoreDBInstanceFromS3

See: newRestoreDBInstanceFromS3 smart constructor.

Constructors

RestoreDBInstanceFromS3'

Fields

engineVersion :: Maybe Text

The version number of the database engine to use. Choose the latest minor version of your database engine. For information about engine versions, see CreateDBInstance, or call DescribeDBEngineVersions.

dbSecurityGroups :: Maybe [Text]

A list of DB security groups to associate with this DB instance.

Default: The default DB security group for the database engine.

deletionProtection :: Maybe Bool

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

storageEncrypted :: Maybe Bool

A value that indicates whether the new DB instance is encrypted or not.

masterUserPassword :: Maybe Text

The password for the master user. The password can include any printable ASCII character except "/", """, or "@".

Constraints: Must contain from 8 to 41 characters.

publiclyAccessible :: Maybe Bool

A value that indicates whether the DB instance is publicly accessible.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

For more information, see CreateDBInstance.

autoMinorVersionUpgrade :: Maybe Bool

A value that indicates whether minor engine upgrades are applied automatically to the DB instance during the maintenance window. By default, minor engine upgrades are not applied automatically.

masterUsername :: Maybe Text

The name for the master user.

Constraints:

Must be 1 to 16 letters or numbers.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

dbSubnetGroupName :: Maybe Text

A DB subnet group to associate with this DB instance.

monitoringRoleArn :: Maybe Text

The ARN for the IAM role that permits RDS to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess. For information on creating a monitoring role, see Setting Up and Enabling Enhanced Monitoring in the Amazon RDS User Guide.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

iops :: Maybe Int

The amount of Provisioned IOPS (input/output operations per second) to allocate initially for the DB instance. For information about valid Iops values, see Amazon RDS Provisioned IOPS Storage to Improve Performance in the Amazon RDS User Guide.

monitoringInterval :: Maybe Int

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance. To disable collecting Enhanced Monitoring metrics, specify 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

Default: 0

processorFeatures :: Maybe [ProcessorFeature]

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

licenseModel :: Maybe Text

The license model for this DB instance. Use general-public-license.

preferredMaintenanceWindow :: Maybe Text

The time range each week during which system maintenance can occur, in Universal Coordinated Time (UTC). For more information, see Amazon RDS Maintenance Window in the Amazon RDS User Guide.

Constraints:

Must be in the format ddd:hh24:mi-ddd:hh24:mi.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred backup window.

Must be at least 30 minutes.

performanceInsightsRetentionPeriod :: Maybe Int

The amount of time, in days, to retain Performance Insights data. Valid values are 7 or 731 (2 years).

maxAllocatedStorage :: Maybe Int

The upper limit in gibibytes (GiB) to which Amazon RDS can automatically scale the storage of the DB instance.

For more information about this setting, including limitations that apply to it, see Managing capacity automatically with Amazon RDS storage autoscaling in the Amazon RDS User Guide.

enablePerformanceInsights :: Maybe Bool

A value that indicates whether to enable Performance Insights for the DB instance.

For more information, see Using Amazon Performance Insights in the Amazon Relational Database Service User Guide.

kmsKeyId :: Maybe Text

The Amazon Web Services KMS key identifier for an encrypted DB instance.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK). To use a CMK in a different Amazon Web Services account, specify the key ARN or alias ARN.

If the StorageEncrypted parameter is enabled, and you do not specify a value for the KmsKeyId parameter, then Amazon RDS will use your default CMK. There is a default CMK for your Amazon Web Services account. Your Amazon Web Services account has a different default CMK for each Amazon Web Services Region.

dbParameterGroupName :: Maybe Text

The name of the DB parameter group to associate with this DB instance.

If you do not specify a value for DBParameterGroupName, then the default DBParameterGroup for the specified DB engine is used.

preferredBackupWindow :: Maybe Text

The time range each day during which automated backups are created if automated backups are enabled. For more information, see Backup window in the Amazon RDS User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

availabilityZone :: Maybe Text

The Availability Zone that the DB instance is created in. For information about Amazon Web Services Regions and Availability Zones, see Regions and Availability Zones in the Amazon RDS User Guide.

Default: A random, system-chosen Availability Zone in the endpoint's Amazon Web Services Region.

Example: us-east-1d

Constraint: The AvailabilityZone parameter can't be specified if the DB instance is a Multi-AZ deployment. The specified Availability Zone must be in the same Amazon Web Services Region as the current endpoint.

backupRetentionPeriod :: Maybe Int

The number of days for which automated backups are retained. Setting this parameter to a positive number enables backups. For more information, see CreateDBInstance.

performanceInsightsKMSKeyId :: Maybe Text

The Amazon Web Services KMS key identifier for encryption of Performance Insights data.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

If you do not specify a value for PerformanceInsightsKMSKeyId, then Amazon RDS uses your default CMK. There is a default CMK for your Amazon Web Services account. Your Amazon Web Services account has a different default CMK for each Amazon Web Services Region.

vpcSecurityGroupIds :: Maybe [Text]

A list of VPC security groups to associate with this DB instance.

multiAZ :: Maybe Bool

A value that indicates whether the DB instance is a Multi-AZ deployment. If the DB instance is a Multi-AZ deployment, you can't set the AvailabilityZone parameter.

s3Prefix :: Maybe Text

The prefix of your Amazon S3 bucket.

allocatedStorage :: Maybe Int

The amount of storage (in gigabytes) to allocate initially for the DB instance. Follow the allocation rules specified in CreateDBInstance.

Be sure to allocate enough memory for your new DB instance so that the restore operation can succeed. You can also allocate additional memory for future growth.

optionGroupName :: Maybe Text

The name of the option group to associate with this DB instance. If this argument is omitted, the default option group for the specified engine is used.

copyTagsToSnapshot :: Maybe Bool

A value that indicates whether to copy all tags from the DB instance to snapshots of the DB instance. By default, tags are not copied.

tags :: Maybe [Tag]

A list of tags to associate with this DB instance. For more information, see Tagging Amazon RDS Resources in the Amazon RDS User Guide.

port :: Maybe Int

The port number on which the database accepts connections.

Type: Integer

Valid Values: 1150-65535

Default: 3306

enableIAMDatabaseAuthentication :: Maybe Bool

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information about IAM database authentication, see IAM Database Authentication for MySQL and PostgreSQL in the Amazon RDS User Guide.

useDefaultProcessorFeatures :: Maybe Bool

A value that indicates whether the DB instance class of the DB instance uses its default processor features.

storageType :: Maybe Text

Specifies the storage type to be associated with the DB instance.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified; otherwise gp2

enableCloudwatchLogsExports :: Maybe [Text]

The list of logs that the restored DB instance is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon RDS User Guide.

dbName :: Maybe Text

The name of the database to create when the DB instance is created. Follow the naming rules specified in CreateDBInstance.

dbInstanceIdentifier :: Text

The DB instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

dbInstanceClass :: Text

The compute and memory capacity of the DB instance, for example, db.m4.large. Not all DB instance classes are available in all Amazon Web Services Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

Importing from Amazon S3 isn't supported on the db.t2.micro DB instance class.

engine :: Text

The name of the database engine to be used for this instance.

Valid Values: mysql

sourceEngine :: Text

The name of the engine of your source database.

Valid Values: mysql

sourceEngineVersion :: Text

The version of the database that the backup files were created from.

MySQL versions 5.6 and 5.7 are supported.

Example: 5.6.40

s3BucketName :: Text

The name of your Amazon S3 bucket that contains your database backup file.

s3IngestionRoleArn :: Text

An Amazon Web Services Identity and Access Management (IAM) role to allow Amazon RDS to access your Amazon S3 bucket.

newRestoreDBInstanceFromS3

Arguments

:: Text
$sel:dbInstanceIdentifier:RestoreDBInstanceFromS3'

-> Text
$sel:dbInstanceClass:RestoreDBInstanceFromS3'

-> Text
$sel:engine:RestoreDBInstanceFromS3'

-> Text
$sel:sourceEngine:RestoreDBInstanceFromS3'

-> Text
$sel:sourceEngineVersion:RestoreDBInstanceFromS3'

-> Text
$sel:s3BucketName:RestoreDBInstanceFromS3'

-> Text
$sel:s3IngestionRoleArn:RestoreDBInstanceFromS3'

-> RestoreDBInstanceFromS3

Create a value of RestoreDBInstanceFromS3 with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_engineVersion - The version number of the database engine to use. Choose the latest minor version of your database engine. For information about engine versions, see CreateDBInstance, or call DescribeDBEngineVersions.

$sel:dbSecurityGroups:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_dbSecurityGroups - A list of DB security groups to associate with this DB instance.

Default: The default DB security group for the database engine.

$sel:deletionProtection:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_deletionProtection - A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

$sel:storageEncrypted:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_storageEncrypted - A value that indicates whether the new DB instance is encrypted or not.

$sel:masterUserPassword:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_masterUserPassword - The password for the master user. The password can include any printable ASCII character except "/", """, or "@".

Constraints: Must contain from 8 to 41 characters.

$sel:publiclyAccessible:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_publiclyAccessible - A value that indicates whether the DB instance is publicly accessible.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

For more information, see CreateDBInstance.

$sel:autoMinorVersionUpgrade:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_autoMinorVersionUpgrade - A value that indicates whether minor engine upgrades are applied automatically to the DB instance during the maintenance window. By default, minor engine upgrades are not applied automatically.

$sel:masterUsername:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_masterUsername - The name for the master user.

Constraints:

Must be 1 to 16 letters or numbers.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

$sel:dbSubnetGroupName:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_dbSubnetGroupName - A DB subnet group to associate with this DB instance.

$sel:monitoringRoleArn:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_monitoringRoleArn - The ARN for the IAM role that permits RDS to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess. For information on creating a monitoring role, see Setting Up and Enabling Enhanced Monitoring in the Amazon RDS User Guide.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

$sel:iops:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_iops - The amount of Provisioned IOPS (input/output operations per second) to allocate initially for the DB instance. For information about valid Iops values, see Amazon RDS Provisioned IOPS Storage to Improve Performance in the Amazon RDS User Guide.

$sel:monitoringInterval:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_monitoringInterval - The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance. To disable collecting Enhanced Monitoring metrics, specify 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

Default: 0

$sel:processorFeatures:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_processorFeatures - The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

$sel:licenseModel:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_licenseModel - The license model for this DB instance. Use general-public-license.

$sel:preferredMaintenanceWindow:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_preferredMaintenanceWindow - The time range each week during which system maintenance can occur, in Universal Coordinated Time (UTC). For more information, see Amazon RDS Maintenance Window in the Amazon RDS User Guide.

Constraints:

Must be in the format ddd:hh24:mi-ddd:hh24:mi.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred backup window.

Must be at least 30 minutes.

$sel:performanceInsightsRetentionPeriod:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_performanceInsightsRetentionPeriod - The amount of time, in days, to retain Performance Insights data. Valid values are 7 or 731 (2 years).

$sel:maxAllocatedStorage:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_maxAllocatedStorage - The upper limit in gibibytes (GiB) to which Amazon RDS can automatically scale the storage of the DB instance.

For more information about this setting, including limitations that apply to it, see Managing capacity automatically with Amazon RDS storage autoscaling in the Amazon RDS User Guide.

$sel:enablePerformanceInsights:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_enablePerformanceInsights - A value that indicates whether to enable Performance Insights for the DB instance.

For more information, see Using Amazon Performance Insights in the Amazon Relational Database Service User Guide.

$sel:kmsKeyId:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_kmsKeyId - The Amazon Web Services KMS key identifier for an encrypted DB instance.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK). To use a CMK in a different Amazon Web Services account, specify the key ARN or alias ARN.

If the StorageEncrypted parameter is enabled, and you do not specify a value for the KmsKeyId parameter, then Amazon RDS will use your default CMK. There is a default CMK for your Amazon Web Services account. Your Amazon Web Services account has a different default CMK for each Amazon Web Services Region.

$sel:dbParameterGroupName:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_dbParameterGroupName - The name of the DB parameter group to associate with this DB instance.

If you do not specify a value for DBParameterGroupName, then the default DBParameterGroup for the specified DB engine is used.

$sel:preferredBackupWindow:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_preferredBackupWindow - The time range each day during which automated backups are created if automated backups are enabled. For more information, see Backup window in the Amazon RDS User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

$sel:availabilityZone:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_availabilityZone - The Availability Zone that the DB instance is created in. For information about Amazon Web Services Regions and Availability Zones, see Regions and Availability Zones in the Amazon RDS User Guide.

Default: A random, system-chosen Availability Zone in the endpoint's Amazon Web Services Region.

Example: us-east-1d

Constraint: The AvailabilityZone parameter can't be specified if the DB instance is a Multi-AZ deployment. The specified Availability Zone must be in the same Amazon Web Services Region as the current endpoint.

$sel:backupRetentionPeriod:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_backupRetentionPeriod - The number of days for which automated backups are retained. Setting this parameter to a positive number enables backups. For more information, see CreateDBInstance.

$sel:performanceInsightsKMSKeyId:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_performanceInsightsKMSKeyId - The Amazon Web Services KMS key identifier for encryption of Performance Insights data.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

If you do not specify a value for PerformanceInsightsKMSKeyId, then Amazon RDS uses your default CMK. There is a default CMK for your Amazon Web Services account. Your Amazon Web Services account has a different default CMK for each Amazon Web Services Region.

$sel:vpcSecurityGroupIds:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_vpcSecurityGroupIds - A list of VPC security groups to associate with this DB instance.

$sel:multiAZ:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_multiAZ - A value that indicates whether the DB instance is a Multi-AZ deployment. If the DB instance is a Multi-AZ deployment, you can't set the AvailabilityZone parameter.

$sel:s3Prefix:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_s3Prefix - The prefix of your Amazon S3 bucket.

$sel:allocatedStorage:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_allocatedStorage - The amount of storage (in gigabytes) to allocate initially for the DB instance. Follow the allocation rules specified in CreateDBInstance.

Be sure to allocate enough memory for your new DB instance so that the restore operation can succeed. You can also allocate additional memory for future growth.

$sel:optionGroupName:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_optionGroupName - The name of the option group to associate with this DB instance. If this argument is omitted, the default option group for the specified engine is used.

$sel:copyTagsToSnapshot:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_copyTagsToSnapshot - A value that indicates whether to copy all tags from the DB instance to snapshots of the DB instance. By default, tags are not copied.

$sel:tags:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_tags - A list of tags to associate with this DB instance. For more information, see Tagging Amazon RDS Resources in the Amazon RDS User Guide.

$sel:port:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_port - The port number on which the database accepts connections.

Type: Integer

Valid Values: 1150-65535

Default: 3306

$sel:enableIAMDatabaseAuthentication:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_enableIAMDatabaseAuthentication - A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information about IAM database authentication, see IAM Database Authentication for MySQL and PostgreSQL in the Amazon RDS User Guide.

$sel:useDefaultProcessorFeatures:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_useDefaultProcessorFeatures - A value that indicates whether the DB instance class of the DB instance uses its default processor features.

$sel:storageType:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_storageType - Specifies the storage type to be associated with the DB instance.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified; otherwise gp2

$sel:enableCloudwatchLogsExports:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_enableCloudwatchLogsExports - The list of logs that the restored DB instance is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon RDS User Guide.

$sel:dbName:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_dbName - The name of the database to create when the DB instance is created. Follow the naming rules specified in CreateDBInstance.

$sel:dbInstanceIdentifier:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_dbInstanceIdentifier - The DB instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

$sel:dbInstanceClass:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_dbInstanceClass - The compute and memory capacity of the DB instance, for example, db.m4.large. Not all DB instance classes are available in all Amazon Web Services Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

Importing from Amazon S3 isn't supported on the db.t2.micro DB instance class.

$sel:engine:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_engine - The name of the database engine to be used for this instance.

Valid Values: mysql

$sel:sourceEngine:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_sourceEngine - The name of the engine of your source database.

Valid Values: mysql

$sel:sourceEngineVersion:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_sourceEngineVersion - The version of the database that the backup files were created from.

MySQL versions 5.6 and 5.7 are supported.

Example: 5.6.40

$sel:s3BucketName:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_s3BucketName - The name of your Amazon S3 bucket that contains your database backup file.

$sel:s3IngestionRoleArn:RestoreDBInstanceFromS3', restoreDBInstanceFromS3_s3IngestionRoleArn - An Amazon Web Services Identity and Access Management (IAM) role to allow Amazon RDS to access your Amazon S3 bucket.

Request Lenses

restoreDBInstanceFromS3_engineVersion :: Lens' RestoreDBInstanceFromS3 (Maybe Text)

The version number of the database engine to use. Choose the latest minor version of your database engine. For information about engine versions, see CreateDBInstance, or call DescribeDBEngineVersions.

restoreDBInstanceFromS3_dbSecurityGroups :: Lens' RestoreDBInstanceFromS3 (Maybe [Text])

A list of DB security groups to associate with this DB instance.

Default: The default DB security group for the database engine.

restoreDBInstanceFromS3_deletionProtection :: Lens' RestoreDBInstanceFromS3 (Maybe Bool)

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

restoreDBInstanceFromS3_storageEncrypted :: Lens' RestoreDBInstanceFromS3 (Maybe Bool)

A value that indicates whether the new DB instance is encrypted or not.

restoreDBInstanceFromS3_masterUserPassword :: Lens' RestoreDBInstanceFromS3 (Maybe Text)

The password for the master user. The password can include any printable ASCII character except "/", """, or "@".

Constraints: Must contain from 8 to 41 characters.

restoreDBInstanceFromS3_publiclyAccessible :: Lens' RestoreDBInstanceFromS3 (Maybe Bool)

A value that indicates whether the DB instance is publicly accessible.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

For more information, see CreateDBInstance.

restoreDBInstanceFromS3_autoMinorVersionUpgrade :: Lens' RestoreDBInstanceFromS3 (Maybe Bool)

A value that indicates whether minor engine upgrades are applied automatically to the DB instance during the maintenance window. By default, minor engine upgrades are not applied automatically.

restoreDBInstanceFromS3_masterUsername :: Lens' RestoreDBInstanceFromS3 (Maybe Text)

The name for the master user.

Constraints:

Must be 1 to 16 letters or numbers.

First character must be a letter.

Can't be a reserved word for the chosen database engine.

restoreDBInstanceFromS3_dbSubnetGroupName :: Lens' RestoreDBInstanceFromS3 (Maybe Text)

A DB subnet group to associate with this DB instance.

restoreDBInstanceFromS3_monitoringRoleArn :: Lens' RestoreDBInstanceFromS3 (Maybe Text)

The ARN for the IAM role that permits RDS to send enhanced monitoring metrics to Amazon CloudWatch Logs. For example, arn:aws:iam:123456789012:role/emaccess. For information on creating a monitoring role, see Setting Up and Enabling Enhanced Monitoring in the Amazon RDS User Guide.

If MonitoringInterval is set to a value other than 0, then you must supply a MonitoringRoleArn value.

restoreDBInstanceFromS3_iops :: Lens' RestoreDBInstanceFromS3 (Maybe Int)

The amount of Provisioned IOPS (input/output operations per second) to allocate initially for the DB instance. For information about valid Iops values, see Amazon RDS Provisioned IOPS Storage to Improve Performance in the Amazon RDS User Guide.

restoreDBInstanceFromS3_monitoringInterval :: Lens' RestoreDBInstanceFromS3 (Maybe Int)

The interval, in seconds, between points when Enhanced Monitoring metrics are collected for the DB instance. To disable collecting Enhanced Monitoring metrics, specify 0.

If MonitoringRoleArn is specified, then you must also set MonitoringInterval to a value other than 0.

Valid Values: 0, 1, 5, 10, 15, 30, 60

Default: 0

restoreDBInstanceFromS3_processorFeatures :: Lens' RestoreDBInstanceFromS3 (Maybe [ProcessorFeature])

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

restoreDBInstanceFromS3_licenseModel :: Lens' RestoreDBInstanceFromS3 (Maybe Text)

The license model for this DB instance. Use general-public-license.

restoreDBInstanceFromS3_preferredMaintenanceWindow :: Lens' RestoreDBInstanceFromS3 (Maybe Text)

The time range each week during which system maintenance can occur, in Universal Coordinated Time (UTC). For more information, see Amazon RDS Maintenance Window in the Amazon RDS User Guide.

Constraints:

Must be in the format ddd:hh24:mi-ddd:hh24:mi.

Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred backup window.

Must be at least 30 minutes.

restoreDBInstanceFromS3_performanceInsightsRetentionPeriod :: Lens' RestoreDBInstanceFromS3 (Maybe Int)

The amount of time, in days, to retain Performance Insights data. Valid values are 7 or 731 (2 years).

restoreDBInstanceFromS3_maxAllocatedStorage :: Lens' RestoreDBInstanceFromS3 (Maybe Int)

The upper limit in gibibytes (GiB) to which Amazon RDS can automatically scale the storage of the DB instance.

For more information about this setting, including limitations that apply to it, see Managing capacity automatically with Amazon RDS storage autoscaling in the Amazon RDS User Guide.

restoreDBInstanceFromS3_enablePerformanceInsights :: Lens' RestoreDBInstanceFromS3 (Maybe Bool)

A value that indicates whether to enable Performance Insights for the DB instance.

For more information, see Using Amazon Performance Insights in the Amazon Relational Database Service User Guide.

restoreDBInstanceFromS3_kmsKeyId :: Lens' RestoreDBInstanceFromS3 (Maybe Text)

The Amazon Web Services KMS key identifier for an encrypted DB instance.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK). To use a CMK in a different Amazon Web Services account, specify the key ARN or alias ARN.

If the StorageEncrypted parameter is enabled, and you do not specify a value for the KmsKeyId parameter, then Amazon RDS will use your default CMK. There is a default CMK for your Amazon Web Services account. Your Amazon Web Services account has a different default CMK for each Amazon Web Services Region.

restoreDBInstanceFromS3_dbParameterGroupName :: Lens' RestoreDBInstanceFromS3 (Maybe Text)

The name of the DB parameter group to associate with this DB instance.

If you do not specify a value for DBParameterGroupName, then the default DBParameterGroup for the specified DB engine is used.

restoreDBInstanceFromS3_preferredBackupWindow :: Lens' RestoreDBInstanceFromS3 (Maybe Text)

The time range each day during which automated backups are created if automated backups are enabled. For more information, see Backup window in the Amazon RDS User Guide.

Constraints:

Must be in the format hh24:mi-hh24:mi.

Must be in Universal Coordinated Time (UTC).

Must not conflict with the preferred maintenance window.

Must be at least 30 minutes.

restoreDBInstanceFromS3_availabilityZone :: Lens' RestoreDBInstanceFromS3 (Maybe Text)

The Availability Zone that the DB instance is created in. For information about Amazon Web Services Regions and Availability Zones, see Regions and Availability Zones in the Amazon RDS User Guide.

Default: A random, system-chosen Availability Zone in the endpoint's Amazon Web Services Region.

Example: us-east-1d

Constraint: The AvailabilityZone parameter can't be specified if the DB instance is a Multi-AZ deployment. The specified Availability Zone must be in the same Amazon Web Services Region as the current endpoint.

restoreDBInstanceFromS3_backupRetentionPeriod :: Lens' RestoreDBInstanceFromS3 (Maybe Int)

The number of days for which automated backups are retained. Setting this parameter to a positive number enables backups. For more information, see CreateDBInstance.

restoreDBInstanceFromS3_performanceInsightsKMSKeyId :: Lens' RestoreDBInstanceFromS3 (Maybe Text)

The Amazon Web Services KMS key identifier for encryption of Performance Insights data.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

If you do not specify a value for PerformanceInsightsKMSKeyId, then Amazon RDS uses your default CMK. There is a default CMK for your Amazon Web Services account. Your Amazon Web Services account has a different default CMK for each Amazon Web Services Region.

restoreDBInstanceFromS3_vpcSecurityGroupIds :: Lens' RestoreDBInstanceFromS3 (Maybe [Text])

A list of VPC security groups to associate with this DB instance.

restoreDBInstanceFromS3_multiAZ :: Lens' RestoreDBInstanceFromS3 (Maybe Bool)

A value that indicates whether the DB instance is a Multi-AZ deployment. If the DB instance is a Multi-AZ deployment, you can't set the AvailabilityZone parameter.

restoreDBInstanceFromS3_s3Prefix :: Lens' RestoreDBInstanceFromS3 (Maybe Text)

The prefix of your Amazon S3 bucket.

restoreDBInstanceFromS3_allocatedStorage :: Lens' RestoreDBInstanceFromS3 (Maybe Int)

The amount of storage (in gigabytes) to allocate initially for the DB instance. Follow the allocation rules specified in CreateDBInstance.

Be sure to allocate enough memory for your new DB instance so that the restore operation can succeed. You can also allocate additional memory for future growth.

restoreDBInstanceFromS3_optionGroupName :: Lens' RestoreDBInstanceFromS3 (Maybe Text)

The name of the option group to associate with this DB instance. If this argument is omitted, the default option group for the specified engine is used.

restoreDBInstanceFromS3_copyTagsToSnapshot :: Lens' RestoreDBInstanceFromS3 (Maybe Bool)

A value that indicates whether to copy all tags from the DB instance to snapshots of the DB instance. By default, tags are not copied.

restoreDBInstanceFromS3_tags :: Lens' RestoreDBInstanceFromS3 (Maybe [Tag])

A list of tags to associate with this DB instance. For more information, see Tagging Amazon RDS Resources in the Amazon RDS User Guide.

restoreDBInstanceFromS3_port :: Lens' RestoreDBInstanceFromS3 (Maybe Int)

The port number on which the database accepts connections.

Type: Integer

Valid Values: 1150-65535

Default: 3306

restoreDBInstanceFromS3_enableIAMDatabaseAuthentication :: Lens' RestoreDBInstanceFromS3 (Maybe Bool)

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information about IAM database authentication, see IAM Database Authentication for MySQL and PostgreSQL in the Amazon RDS User Guide.

restoreDBInstanceFromS3_useDefaultProcessorFeatures :: Lens' RestoreDBInstanceFromS3 (Maybe Bool)

A value that indicates whether the DB instance class of the DB instance uses its default processor features.

restoreDBInstanceFromS3_storageType :: Lens' RestoreDBInstanceFromS3 (Maybe Text)

Specifies the storage type to be associated with the DB instance.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified; otherwise gp2

restoreDBInstanceFromS3_enableCloudwatchLogsExports :: Lens' RestoreDBInstanceFromS3 (Maybe [Text])

The list of logs that the restored DB instance is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon RDS User Guide.

restoreDBInstanceFromS3_dbName :: Lens' RestoreDBInstanceFromS3 (Maybe Text)

The name of the database to create when the DB instance is created. Follow the naming rules specified in CreateDBInstance.

restoreDBInstanceFromS3_dbInstanceIdentifier :: Lens' RestoreDBInstanceFromS3 Text

The DB instance identifier. This parameter is stored as a lowercase string.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

Example: mydbinstance

restoreDBInstanceFromS3_dbInstanceClass :: Lens' RestoreDBInstanceFromS3 Text

The compute and memory capacity of the DB instance, for example, db.m4.large. Not all DB instance classes are available in all Amazon Web Services Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

Importing from Amazon S3 isn't supported on the db.t2.micro DB instance class.

restoreDBInstanceFromS3_engine :: Lens' RestoreDBInstanceFromS3 Text

The name of the database engine to be used for this instance.

Valid Values: mysql

restoreDBInstanceFromS3_sourceEngine :: Lens' RestoreDBInstanceFromS3 Text

The name of the engine of your source database.

Valid Values: mysql

restoreDBInstanceFromS3_sourceEngineVersion :: Lens' RestoreDBInstanceFromS3 Text

The version of the database that the backup files were created from.

MySQL versions 5.6 and 5.7 are supported.

Example: 5.6.40

restoreDBInstanceFromS3_s3BucketName :: Lens' RestoreDBInstanceFromS3 Text

The name of your Amazon S3 bucket that contains your database backup file.

restoreDBInstanceFromS3_s3IngestionRoleArn :: Lens' RestoreDBInstanceFromS3 Text

An Amazon Web Services Identity and Access Management (IAM) role to allow Amazon RDS to access your Amazon S3 bucket.

Destructuring the Response

data RestoreDBInstanceFromS3Response

See: newRestoreDBInstanceFromS3Response smart constructor.

Constructors

RestoreDBInstanceFromS3Response'

Fields

dbInstance :: Maybe DBInstance

httpStatus :: Int

The response's http status code.

newRestoreDBInstanceFromS3Response

Arguments

:: Int
$sel:httpStatus:RestoreDBInstanceFromS3Response'

-> RestoreDBInstanceFromS3Response

Create a value of RestoreDBInstanceFromS3Response with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbInstance:RestoreDBInstanceFromS3Response', restoreDBInstanceFromS3Response_dbInstance - Undocumented member.

$sel:httpStatus:RestoreDBInstanceFromS3Response', restoreDBInstanceFromS3Response_httpStatus - The response's http status code.

Response Lenses

restoreDBInstanceFromS3Response_dbInstance :: Lens' RestoreDBInstanceFromS3Response (Maybe DBInstance)

Undocumented member.

restoreDBInstanceFromS3Response_httpStatus :: Lens' RestoreDBInstanceFromS3Response Int

The response's http status code.