Network.AWS.RDS.RestoreDBInstanceToPointInTime

Description

Restores a DB instance to an arbitrary point in time. You can restore to any point in time before the time identified by the LatestRestorableTime property. You can restore to a point up to the number of days specified by the BackupRetentionPeriod property.

The target database is created with most of the original configuration, but in a system-selected Availability Zone, with the default security group, the default subnet group, and the default DB parameter group. By default, the new DB instance is created as a single-AZ deployment except when the instance is a SQL Server instance that has an option group that is associated with mirroring; in this case, the instance becomes a mirrored deployment and not a single-AZ deployment.

This command doesn't apply to Aurora MySQL and Aurora PostgreSQL. For Aurora, use RestoreDBClusterToPointInTime.

Creating a Request

data RestoreDBInstanceToPointInTime

See: newRestoreDBInstanceToPointInTime smart constructor.

Constructors

RestoreDBInstanceToPointInTime'

Fields

deletionProtection :: Maybe Bool

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

useLatestRestorableTime :: Maybe Bool

A value that indicates whether the DB instance is restored from the latest backup time. By default, the DB instance isn't restored from the latest backup time.

Constraints: Can't be specified if the RestoreTime parameter is provided.

publiclyAccessible :: Maybe Bool

A value that indicates whether the DB instance is publicly accessible.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

For more information, see CreateDBInstance.

autoMinorVersionUpgrade :: Maybe Bool

A value that indicates whether minor version upgrades are applied automatically to the DB instance during the maintenance window.

dbSubnetGroupName :: Maybe Text

The DB subnet group name to use for the new instance.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

restoreTime :: Maybe ISO8601

The date and time to restore from.

Valid Values: Value must be a time in Universal Coordinated Time (UTC) format

Constraints:

Must be before the latest restorable time for the DB instance

Can't be specified if the UseLatestRestorableTime parameter is enabled

Example: 2009-09-07T23:45:00Z

iops :: Maybe Int

The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance.

Constraints: Must be an integer greater than 1000.

SQL Server

Setting the IOPS value for the SQL Server database engine isn't supported.

domain :: Maybe Text

Specify the Active Directory directory ID to restore the DB instance in. The domain must be created prior to this operation. Currently, only MySQL, Microsoft SQL Server, Oracle, and PostgreSQL DB instances can be created in an Active Directory Domain.

For more information, see Kerberos Authentication in the Amazon RDS User Guide.

enableCustomerOwnedIp :: Maybe Bool

A value that indicates whether to enable a customer-owned IP address (CoIP) for an RDS on Outposts DB instance.

A CoIP provides local or external connectivity to resources in your Outpost subnets through your on-premises network. For some use cases, a CoIP can provide lower latency for connections to the DB instance from outside of its virtual private cloud (VPC) on your local network.

For more information about RDS on Outposts, see Working with Amazon RDS on Amazon Web Services Outposts in the Amazon RDS User Guide.

For more information about CoIPs, see Customer-owned IP addresses in the Amazon Web Services Outposts User Guide.

engine :: Maybe Text

The database engine to use for the new instance.

Default: The same as source

Constraint: Must be compatible with the engine of the source

Valid Values:

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

tdeCredentialPassword :: Maybe Text

The password for the given ARN from the key store in order to access the device.

sourceDBInstanceIdentifier :: Maybe Text

The identifier of the source DB instance from which to restore.

Constraints:

Must match the identifier of an existing DB instance.

processorFeatures :: Maybe [ProcessorFeature]

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

dbInstanceClass :: Maybe Text

The compute and memory capacity of the Amazon RDS DB instance, for example, db.m4.large. Not all DB instance classes are available in all Amazon Web Services Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

Default: The same DBInstanceClass as the original DB instance.

licenseModel :: Maybe Text

License model information for the restored DB instance.

Default: Same as source.

Valid values: license-included | bring-your-own-license | general-public-license

sourceDBInstanceAutomatedBackupsArn :: Maybe Text

The Amazon Resource Name (ARN) of the replicated automated backups from which to restore, for example, arn:aws:rds:useast-1:123456789012:auto-backup:ab-L2IJCEXJP7XQ7HOJ4SIEXAMPLE.

maxAllocatedStorage :: Maybe Int

The upper limit in gibibytes (GiB) to which Amazon RDS can automatically scale the storage of the DB instance.

For more information about this setting, including limitations that apply to it, see Managing capacity automatically with Amazon RDS storage autoscaling in the Amazon RDS User Guide.

dbParameterGroupName :: Maybe Text

The name of the DB parameter group to associate with this DB instance.

If you do not specify a value for DBParameterGroupName, then the default DBParameterGroup for the specified DB engine is used.

Constraints:

If supplied, must match the name of an existing DBParameterGroup.

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

availabilityZone :: Maybe Text

The Availability Zone (AZ) where the DB instance will be created.

Default: A random, system-chosen Availability Zone.

Constraint: You can't specify the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

Example: us-east-1a

vpcSecurityGroupIds :: Maybe [Text]

A list of EC2 VPC security groups to associate with this DB instance.

Default: The default EC2 VPC security group for the DB subnet group's VPC.

multiAZ :: Maybe Bool

A value that indicates whether the DB instance is a Multi-AZ deployment.

Constraint: You can't specify the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

sourceDbiResourceId :: Maybe Text

The resource ID of the source DB instance from which to restore.

optionGroupName :: Maybe Text

The name of the option group to be used for the restored DB instance.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, can't be removed from an option group, and that option group can't be removed from a DB instance once it is associated with a DB instance

copyTagsToSnapshot :: Maybe Bool

A value that indicates whether to copy all tags from the restored DB instance to snapshots of the DB instance. By default, tags are not copied.

tdeCredentialArn :: Maybe Text

The ARN from the key store with which to associate the instance for TDE encryption.

domainIAMRoleName :: Maybe Text

Specify the name of the IAM role to be used when making API calls to the Directory Service.

tags :: Maybe [Tag]

port :: Maybe Int

The port number on which the database accepts connections.

Constraints: Value must be 1150-65535

Default: The same port as the original DB instance.

enableIAMDatabaseAuthentication :: Maybe Bool

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information about IAM database authentication, see IAM Database Authentication for MySQL and PostgreSQL in the Amazon RDS User Guide.

useDefaultProcessorFeatures :: Maybe Bool

A value that indicates whether the DB instance class of the DB instance uses its default processor features.

storageType :: Maybe Text

Specifies the storage type to be associated with the DB instance.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified, otherwise gp2

enableCloudwatchLogsExports :: Maybe [Text]

The list of logs that the restored DB instance is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon RDS User Guide.

dbName :: Maybe Text

The database name for the restored DB instance.

This parameter isn't used for the MySQL or MariaDB engines.

targetDBInstanceIdentifier :: Text

The name of the new DB instance to be created.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

newRestoreDBInstanceToPointInTime

Arguments

:: Text
$sel:targetDBInstanceIdentifier:RestoreDBInstanceToPointInTime'

-> RestoreDBInstanceToPointInTime

Create a value of RestoreDBInstanceToPointInTime with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:deletionProtection:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_deletionProtection - A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

$sel:useLatestRestorableTime:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_useLatestRestorableTime - A value that indicates whether the DB instance is restored from the latest backup time. By default, the DB instance isn't restored from the latest backup time.

Constraints: Can't be specified if the RestoreTime parameter is provided.

$sel:publiclyAccessible:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_publiclyAccessible - A value that indicates whether the DB instance is publicly accessible.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

For more information, see CreateDBInstance.

$sel:autoMinorVersionUpgrade:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_autoMinorVersionUpgrade - A value that indicates whether minor version upgrades are applied automatically to the DB instance during the maintenance window.

$sel:dbSubnetGroupName:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_dbSubnetGroupName - The DB subnet group name to use for the new instance.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

$sel:restoreTime:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_restoreTime - The date and time to restore from.

Valid Values: Value must be a time in Universal Coordinated Time (UTC) format

Constraints:

Must be before the latest restorable time for the DB instance

Can't be specified if the UseLatestRestorableTime parameter is enabled

Example: 2009-09-07T23:45:00Z

$sel:iops:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_iops - The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance.

Constraints: Must be an integer greater than 1000.

SQL Server

Setting the IOPS value for the SQL Server database engine isn't supported.

$sel:domain:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_domain - Specify the Active Directory directory ID to restore the DB instance in. The domain must be created prior to this operation. Currently, only MySQL, Microsoft SQL Server, Oracle, and PostgreSQL DB instances can be created in an Active Directory Domain.

For more information, see Kerberos Authentication in the Amazon RDS User Guide.

$sel:enableCustomerOwnedIp:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_enableCustomerOwnedIp - A value that indicates whether to enable a customer-owned IP address (CoIP) for an RDS on Outposts DB instance.

A CoIP provides local or external connectivity to resources in your Outpost subnets through your on-premises network. For some use cases, a CoIP can provide lower latency for connections to the DB instance from outside of its virtual private cloud (VPC) on your local network.

For more information about RDS on Outposts, see Working with Amazon RDS on Amazon Web Services Outposts in the Amazon RDS User Guide.

For more information about CoIPs, see Customer-owned IP addresses in the Amazon Web Services Outposts User Guide.

$sel:engine:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_engine - The database engine to use for the new instance.

Default: The same as source

Constraint: Must be compatible with the engine of the source

Valid Values:

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

$sel:tdeCredentialPassword:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_tdeCredentialPassword - The password for the given ARN from the key store in order to access the device.

$sel:sourceDBInstanceIdentifier:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_sourceDBInstanceIdentifier - The identifier of the source DB instance from which to restore.

Constraints:

Must match the identifier of an existing DB instance.

$sel:processorFeatures:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_processorFeatures - The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

$sel:dbInstanceClass:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_dbInstanceClass - The compute and memory capacity of the Amazon RDS DB instance, for example, db.m4.large. Not all DB instance classes are available in all Amazon Web Services Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

Default: The same DBInstanceClass as the original DB instance.

$sel:licenseModel:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_licenseModel - License model information for the restored DB instance.

Default: Same as source.

Valid values: license-included | bring-your-own-license | general-public-license

$sel:sourceDBInstanceAutomatedBackupsArn:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_sourceDBInstanceAutomatedBackupsArn - The Amazon Resource Name (ARN) of the replicated automated backups from which to restore, for example, arn:aws:rds:useast-1:123456789012:auto-backup:ab-L2IJCEXJP7XQ7HOJ4SIEXAMPLE.

$sel:maxAllocatedStorage:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_maxAllocatedStorage - The upper limit in gibibytes (GiB) to which Amazon RDS can automatically scale the storage of the DB instance.

For more information about this setting, including limitations that apply to it, see Managing capacity automatically with Amazon RDS storage autoscaling in the Amazon RDS User Guide.

$sel:dbParameterGroupName:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_dbParameterGroupName - The name of the DB parameter group to associate with this DB instance.

If you do not specify a value for DBParameterGroupName, then the default DBParameterGroup for the specified DB engine is used.

Constraints:

If supplied, must match the name of an existing DBParameterGroup.

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

$sel:availabilityZone:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_availabilityZone - The Availability Zone (AZ) where the DB instance will be created.

Default: A random, system-chosen Availability Zone.

Constraint: You can't specify the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

Example: us-east-1a

$sel:vpcSecurityGroupIds:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_vpcSecurityGroupIds - A list of EC2 VPC security groups to associate with this DB instance.

Default: The default EC2 VPC security group for the DB subnet group's VPC.

$sel:multiAZ:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_multiAZ - A value that indicates whether the DB instance is a Multi-AZ deployment.

Constraint: You can't specify the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

$sel:sourceDbiResourceId:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_sourceDbiResourceId - The resource ID of the source DB instance from which to restore.

$sel:optionGroupName:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_optionGroupName - The name of the option group to be used for the restored DB instance.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, can't be removed from an option group, and that option group can't be removed from a DB instance once it is associated with a DB instance

$sel:copyTagsToSnapshot:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_copyTagsToSnapshot - A value that indicates whether to copy all tags from the restored DB instance to snapshots of the DB instance. By default, tags are not copied.

$sel:tdeCredentialArn:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_tdeCredentialArn - The ARN from the key store with which to associate the instance for TDE encryption.

$sel:domainIAMRoleName:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_domainIAMRoleName - Specify the name of the IAM role to be used when making API calls to the Directory Service.

$sel:tags:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_tags - Undocumented member.

$sel:port:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_port - The port number on which the database accepts connections.

Constraints: Value must be 1150-65535

Default: The same port as the original DB instance.

$sel:enableIAMDatabaseAuthentication:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_enableIAMDatabaseAuthentication - A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information about IAM database authentication, see IAM Database Authentication for MySQL and PostgreSQL in the Amazon RDS User Guide.

$sel:useDefaultProcessorFeatures:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_useDefaultProcessorFeatures - A value that indicates whether the DB instance class of the DB instance uses its default processor features.

$sel:storageType:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_storageType - Specifies the storage type to be associated with the DB instance.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified, otherwise gp2

$sel:enableCloudwatchLogsExports:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_enableCloudwatchLogsExports - The list of logs that the restored DB instance is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon RDS User Guide.

$sel:dbName:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_dbName - The database name for the restored DB instance.

This parameter isn't used for the MySQL or MariaDB engines.

$sel:targetDBInstanceIdentifier:RestoreDBInstanceToPointInTime', restoreDBInstanceToPointInTime_targetDBInstanceIdentifier - The name of the new DB instance to be created.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Request Lenses

restoreDBInstanceToPointInTime_deletionProtection :: Lens' RestoreDBInstanceToPointInTime (Maybe Bool)

A value that indicates whether the DB instance has deletion protection enabled. The database can't be deleted when deletion protection is enabled. By default, deletion protection is disabled. For more information, see Deleting a DB Instance.

restoreDBInstanceToPointInTime_useLatestRestorableTime :: Lens' RestoreDBInstanceToPointInTime (Maybe Bool)

A value that indicates whether the DB instance is restored from the latest backup time. By default, the DB instance isn't restored from the latest backup time.

Constraints: Can't be specified if the RestoreTime parameter is provided.

restoreDBInstanceToPointInTime_publiclyAccessible :: Lens' RestoreDBInstanceToPointInTime (Maybe Bool)

A value that indicates whether the DB instance is publicly accessible.

When the DB instance is publicly accessible, its DNS endpoint resolves to the private IP address from within the DB instance's VPC, and to the public IP address from outside of the DB instance's VPC. Access to the DB instance is ultimately controlled by the security group it uses, and that public access is not permitted if the security group assigned to the DB instance doesn't permit it.

When the DB instance isn't publicly accessible, it is an internal DB instance with a DNS name that resolves to a private IP address.

For more information, see CreateDBInstance.

restoreDBInstanceToPointInTime_autoMinorVersionUpgrade :: Lens' RestoreDBInstanceToPointInTime (Maybe Bool)

A value that indicates whether minor version upgrades are applied automatically to the DB instance during the maintenance window.

restoreDBInstanceToPointInTime_dbSubnetGroupName :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The DB subnet group name to use for the new instance.

Constraints: If supplied, must match the name of an existing DBSubnetGroup.

Example: mySubnetgroup

restoreDBInstanceToPointInTime_restoreTime :: Lens' RestoreDBInstanceToPointInTime (Maybe UTCTime)

The date and time to restore from.

Valid Values: Value must be a time in Universal Coordinated Time (UTC) format

Constraints:

Must be before the latest restorable time for the DB instance

Can't be specified if the UseLatestRestorableTime parameter is enabled

Example: 2009-09-07T23:45:00Z

restoreDBInstanceToPointInTime_iops :: Lens' RestoreDBInstanceToPointInTime (Maybe Int)

The amount of Provisioned IOPS (input/output operations per second) to be initially allocated for the DB instance.

Constraints: Must be an integer greater than 1000.

SQL Server

Setting the IOPS value for the SQL Server database engine isn't supported.

restoreDBInstanceToPointInTime_domain :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

Specify the Active Directory directory ID to restore the DB instance in. The domain must be created prior to this operation. Currently, only MySQL, Microsoft SQL Server, Oracle, and PostgreSQL DB instances can be created in an Active Directory Domain.

For more information, see Kerberos Authentication in the Amazon RDS User Guide.

restoreDBInstanceToPointInTime_enableCustomerOwnedIp :: Lens' RestoreDBInstanceToPointInTime (Maybe Bool)

A value that indicates whether to enable a customer-owned IP address (CoIP) for an RDS on Outposts DB instance.

A CoIP provides local or external connectivity to resources in your Outpost subnets through your on-premises network. For some use cases, a CoIP can provide lower latency for connections to the DB instance from outside of its virtual private cloud (VPC) on your local network.

For more information about RDS on Outposts, see Working with Amazon RDS on Amazon Web Services Outposts in the Amazon RDS User Guide.

For more information about CoIPs, see Customer-owned IP addresses in the Amazon Web Services Outposts User Guide.

restoreDBInstanceToPointInTime_engine :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The database engine to use for the new instance.

Default: The same as source

Constraint: Must be compatible with the engine of the source

Valid Values:

mariadb

mysql

oracle-ee

oracle-ee-cdb

oracle-se2

oracle-se2-cdb

postgres

sqlserver-ee

sqlserver-se

sqlserver-ex

sqlserver-web

restoreDBInstanceToPointInTime_tdeCredentialPassword :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The password for the given ARN from the key store in order to access the device.

restoreDBInstanceToPointInTime_sourceDBInstanceIdentifier :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The identifier of the source DB instance from which to restore.

Constraints:

Must match the identifier of an existing DB instance.

restoreDBInstanceToPointInTime_processorFeatures :: Lens' RestoreDBInstanceToPointInTime (Maybe [ProcessorFeature])

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

restoreDBInstanceToPointInTime_dbInstanceClass :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The compute and memory capacity of the Amazon RDS DB instance, for example, db.m4.large. Not all DB instance classes are available in all Amazon Web Services Regions, or for all database engines. For the full list of DB instance classes, and availability for your engine, see DB Instance Class in the Amazon RDS User Guide.

Default: The same DBInstanceClass as the original DB instance.

restoreDBInstanceToPointInTime_licenseModel :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

License model information for the restored DB instance.

Default: Same as source.

Valid values: license-included | bring-your-own-license | general-public-license

restoreDBInstanceToPointInTime_sourceDBInstanceAutomatedBackupsArn :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The Amazon Resource Name (ARN) of the replicated automated backups from which to restore, for example, arn:aws:rds:useast-1:123456789012:auto-backup:ab-L2IJCEXJP7XQ7HOJ4SIEXAMPLE.

restoreDBInstanceToPointInTime_maxAllocatedStorage :: Lens' RestoreDBInstanceToPointInTime (Maybe Int)

The upper limit in gibibytes (GiB) to which Amazon RDS can automatically scale the storage of the DB instance.

For more information about this setting, including limitations that apply to it, see Managing capacity automatically with Amazon RDS storage autoscaling in the Amazon RDS User Guide.

restoreDBInstanceToPointInTime_dbParameterGroupName :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The name of the DB parameter group to associate with this DB instance.

If you do not specify a value for DBParameterGroupName, then the default DBParameterGroup for the specified DB engine is used.

Constraints:

If supplied, must match the name of an existing DBParameterGroup.

Must be 1 to 255 letters, numbers, or hyphens.

First character must be a letter.

Can't end with a hyphen or contain two consecutive hyphens.

restoreDBInstanceToPointInTime_availabilityZone :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The Availability Zone (AZ) where the DB instance will be created.

Default: A random, system-chosen Availability Zone.

Constraint: You can't specify the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

Example: us-east-1a

restoreDBInstanceToPointInTime_vpcSecurityGroupIds :: Lens' RestoreDBInstanceToPointInTime (Maybe [Text])

A list of EC2 VPC security groups to associate with this DB instance.

Default: The default EC2 VPC security group for the DB subnet group's VPC.

restoreDBInstanceToPointInTime_multiAZ :: Lens' RestoreDBInstanceToPointInTime (Maybe Bool)

A value that indicates whether the DB instance is a Multi-AZ deployment.

Constraint: You can't specify the AvailabilityZone parameter if the DB instance is a Multi-AZ deployment.

restoreDBInstanceToPointInTime_sourceDbiResourceId :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The resource ID of the source DB instance from which to restore.

restoreDBInstanceToPointInTime_optionGroupName :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The name of the option group to be used for the restored DB instance.

Permanent options, such as the TDE option for Oracle Advanced Security TDE, can't be removed from an option group, and that option group can't be removed from a DB instance once it is associated with a DB instance

restoreDBInstanceToPointInTime_copyTagsToSnapshot :: Lens' RestoreDBInstanceToPointInTime (Maybe Bool)

A value that indicates whether to copy all tags from the restored DB instance to snapshots of the DB instance. By default, tags are not copied.

restoreDBInstanceToPointInTime_tdeCredentialArn :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The ARN from the key store with which to associate the instance for TDE encryption.

restoreDBInstanceToPointInTime_domainIAMRoleName :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

Specify the name of the IAM role to be used when making API calls to the Directory Service.

restoreDBInstanceToPointInTime_tags :: Lens' RestoreDBInstanceToPointInTime (Maybe [Tag])

Undocumented member.

restoreDBInstanceToPointInTime_port :: Lens' RestoreDBInstanceToPointInTime (Maybe Int)

The port number on which the database accepts connections.

Constraints: Value must be 1150-65535

Default: The same port as the original DB instance.

restoreDBInstanceToPointInTime_enableIAMDatabaseAuthentication :: Lens' RestoreDBInstanceToPointInTime (Maybe Bool)

A value that indicates whether to enable mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts. By default, mapping is disabled.

For more information about IAM database authentication, see IAM Database Authentication for MySQL and PostgreSQL in the Amazon RDS User Guide.

restoreDBInstanceToPointInTime_useDefaultProcessorFeatures :: Lens' RestoreDBInstanceToPointInTime (Maybe Bool)

A value that indicates whether the DB instance class of the DB instance uses its default processor features.

restoreDBInstanceToPointInTime_storageType :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

Specifies the storage type to be associated with the DB instance.

Valid values: standard | gp2 | io1

If you specify io1, you must also include a value for the Iops parameter.

Default: io1 if the Iops parameter is specified, otherwise gp2

restoreDBInstanceToPointInTime_enableCloudwatchLogsExports :: Lens' RestoreDBInstanceToPointInTime (Maybe [Text])

The list of logs that the restored DB instance is to export to CloudWatch Logs. The values in the list depend on the DB engine being used. For more information, see Publishing Database Logs to Amazon CloudWatch Logs in the Amazon RDS User Guide.

restoreDBInstanceToPointInTime_dbName :: Lens' RestoreDBInstanceToPointInTime (Maybe Text)

The database name for the restored DB instance.

This parameter isn't used for the MySQL or MariaDB engines.

restoreDBInstanceToPointInTime_targetDBInstanceIdentifier :: Lens' RestoreDBInstanceToPointInTime Text

The name of the new DB instance to be created.

Constraints:

Must contain from 1 to 63 letters, numbers, or hyphens

First character must be a letter

Can't end with a hyphen or contain two consecutive hyphens

Destructuring the Response

data RestoreDBInstanceToPointInTimeResponse

See: newRestoreDBInstanceToPointInTimeResponse smart constructor.

Constructors

RestoreDBInstanceToPointInTimeResponse'

Fields

dbInstance :: Maybe DBInstance

httpStatus :: Int

The response's http status code.

newRestoreDBInstanceToPointInTimeResponse

Arguments

:: Int
$sel:httpStatus:RestoreDBInstanceToPointInTimeResponse'

-> RestoreDBInstanceToPointInTimeResponse

Create a value of RestoreDBInstanceToPointInTimeResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbInstance:RestoreDBInstanceToPointInTimeResponse', restoreDBInstanceToPointInTimeResponse_dbInstance - Undocumented member.

$sel:httpStatus:RestoreDBInstanceToPointInTimeResponse', restoreDBInstanceToPointInTimeResponse_httpStatus - The response's http status code.

Response Lenses

restoreDBInstanceToPointInTimeResponse_dbInstance :: Lens' RestoreDBInstanceToPointInTimeResponse (Maybe DBInstance)

Undocumented member.

restoreDBInstanceToPointInTimeResponse_httpStatus :: Lens' RestoreDBInstanceToPointInTimeResponse Int

The response's http status code.