Network.AWS.RDS.RevokeDBSecurityGroupIngress

Description

Revokes ingress from a DBSecurityGroup for previously authorized IP ranges or EC2 or VPC security groups. Required parameters for this API are one of CIDRIP, EC2SecurityGroupId for VPC, or (EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId).

Creating a Request

data RevokeDBSecurityGroupIngress

See: newRevokeDBSecurityGroupIngress smart constructor.

Constructors

RevokeDBSecurityGroupIngress'

Fields

cidrip :: Maybe Text

The IP range to revoke access from. Must be a valid CIDR range. If CIDRIP is specified, EC2SecurityGroupName, EC2SecurityGroupId and EC2SecurityGroupOwnerId can't be provided.

eC2SecurityGroupOwnerId :: Maybe Text

The Amazon Web Services account number of the owner of the EC2 security group specified in the EC2SecurityGroupName parameter. The Amazon Web Services access key ID isn't an acceptable value. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

eC2SecurityGroupId :: Maybe Text

The id of the EC2 security group to revoke access from. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

eC2SecurityGroupName :: Maybe Text

The name of the EC2 security group to revoke access from. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

dbSecurityGroupName :: Text

The name of the DB security group to revoke ingress from.

newRevokeDBSecurityGroupIngress

Arguments

:: Text
$sel:dbSecurityGroupName:RevokeDBSecurityGroupIngress'

-> RevokeDBSecurityGroupIngress

Create a value of RevokeDBSecurityGroupIngress with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cidrip:RevokeDBSecurityGroupIngress', revokeDBSecurityGroupIngress_cidrip - The IP range to revoke access from. Must be a valid CIDR range. If CIDRIP is specified, EC2SecurityGroupName, EC2SecurityGroupId and EC2SecurityGroupOwnerId can't be provided.

$sel:eC2SecurityGroupOwnerId:RevokeDBSecurityGroupIngress', revokeDBSecurityGroupIngress_eC2SecurityGroupOwnerId - The Amazon Web Services account number of the owner of the EC2 security group specified in the EC2SecurityGroupName parameter. The Amazon Web Services access key ID isn't an acceptable value. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

$sel:eC2SecurityGroupId:RevokeDBSecurityGroupIngress', revokeDBSecurityGroupIngress_eC2SecurityGroupId - The id of the EC2 security group to revoke access from. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

$sel:eC2SecurityGroupName:RevokeDBSecurityGroupIngress', revokeDBSecurityGroupIngress_eC2SecurityGroupName - The name of the EC2 security group to revoke access from. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

$sel:dbSecurityGroupName:RevokeDBSecurityGroupIngress', revokeDBSecurityGroupIngress_dbSecurityGroupName - The name of the DB security group to revoke ingress from.

Request Lenses

revokeDBSecurityGroupIngress_cidrip :: Lens' RevokeDBSecurityGroupIngress (Maybe Text)

The IP range to revoke access from. Must be a valid CIDR range. If CIDRIP is specified, EC2SecurityGroupName, EC2SecurityGroupId and EC2SecurityGroupOwnerId can't be provided.

revokeDBSecurityGroupIngress_eC2SecurityGroupOwnerId :: Lens' RevokeDBSecurityGroupIngress (Maybe Text)

The Amazon Web Services account number of the owner of the EC2 security group specified in the EC2SecurityGroupName parameter. The Amazon Web Services access key ID isn't an acceptable value. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

revokeDBSecurityGroupIngress_eC2SecurityGroupId :: Lens' RevokeDBSecurityGroupIngress (Maybe Text)

The id of the EC2 security group to revoke access from. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

revokeDBSecurityGroupIngress_eC2SecurityGroupName :: Lens' RevokeDBSecurityGroupIngress (Maybe Text)

The name of the EC2 security group to revoke access from. For VPC DB security groups, EC2SecurityGroupId must be provided. Otherwise, EC2SecurityGroupOwnerId and either EC2SecurityGroupName or EC2SecurityGroupId must be provided.

revokeDBSecurityGroupIngress_dbSecurityGroupName :: Lens' RevokeDBSecurityGroupIngress Text

The name of the DB security group to revoke ingress from.

Destructuring the Response

data RevokeDBSecurityGroupIngressResponse

See: newRevokeDBSecurityGroupIngressResponse smart constructor.

Constructors

RevokeDBSecurityGroupIngressResponse'

Fields

dbSecurityGroup :: Maybe DBSecurityGroup

httpStatus :: Int

The response's http status code.

newRevokeDBSecurityGroupIngressResponse

Arguments

:: Int
$sel:httpStatus:RevokeDBSecurityGroupIngressResponse'

-> RevokeDBSecurityGroupIngressResponse

Create a value of RevokeDBSecurityGroupIngressResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbSecurityGroup:RevokeDBSecurityGroupIngressResponse', revokeDBSecurityGroupIngressResponse_dbSecurityGroup - Undocumented member.

$sel:httpStatus:RevokeDBSecurityGroupIngressResponse', revokeDBSecurityGroupIngressResponse_httpStatus - The response's http status code.

Response Lenses

revokeDBSecurityGroupIngressResponse_dbSecurityGroup :: Lens' RevokeDBSecurityGroupIngressResponse (Maybe DBSecurityGroup)

Undocumented member.

revokeDBSecurityGroupIngressResponse_httpStatus :: Lens' RevokeDBSecurityGroupIngressResponse Int

The response's http status code.