Network.AWS.RDS.StartDBInstance

Description

Starts an Amazon RDS DB instance that was stopped using the Amazon Web Services console, the stop-db-instance CLI command, or the StopDBInstance action.

For more information, see Starting an Amazon RDS DB instance That Was Previously Stopped in the Amazon RDS User Guide.

This command doesn't apply to Aurora MySQL and Aurora PostgreSQL. For Aurora DB clusters, use StartDBCluster instead.

Creating a Request

data StartDBInstance

See: newStartDBInstance smart constructor.

Constructors

StartDBInstance'

Fields

dbInstanceIdentifier :: Text

The user-supplied instance identifier.

newStartDBInstance

Arguments

:: Text
$sel:dbInstanceIdentifier:StartDBInstance'

-> StartDBInstance

Create a value of StartDBInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbInstanceIdentifier:StartDBInstance', startDBInstance_dbInstanceIdentifier - The user-supplied instance identifier.

Request Lenses

startDBInstance_dbInstanceIdentifier :: Lens' StartDBInstance Text

The user-supplied instance identifier.

Destructuring the Response

data StartDBInstanceResponse

See: newStartDBInstanceResponse smart constructor.

Constructors

StartDBInstanceResponse'

Fields

dbInstance :: Maybe DBInstance

httpStatus :: Int

The response's http status code.

newStartDBInstanceResponse

Arguments

:: Int
$sel:httpStatus:StartDBInstanceResponse'

-> StartDBInstanceResponse

Create a value of StartDBInstanceResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbInstance:StartDBInstanceResponse', startDBInstanceResponse_dbInstance - Undocumented member.

$sel:httpStatus:StartDBInstanceResponse', startDBInstanceResponse_httpStatus - The response's http status code.

Response Lenses

startDBInstanceResponse_dbInstance :: Lens' StartDBInstanceResponse (Maybe DBInstance)

Undocumented member.

startDBInstanceResponse_httpStatus :: Lens' StartDBInstanceResponse Int

The response's http status code.