Network.AWS.RDS.StopActivityStream

Description

Stops a database activity stream that was started using the Amazon Web Services console, the start-activity-stream CLI command, or the StartActivityStream action.

For more information, see Database Activity Streams in the Amazon Aurora User Guide.

Creating a Request

data StopActivityStream

See: newStopActivityStream smart constructor.

Constructors

StopActivityStream'

Fields

applyImmediately :: Maybe Bool

Specifies whether or not the database activity stream is to stop as soon as possible, regardless of the maintenance window for the database.

resourceArn :: Text

The Amazon Resource Name (ARN) of the DB cluster for the database activity stream. For example, arn:aws:rds:us-east-1:12345667890:cluster:das-cluster.

newStopActivityStream

Arguments

:: Text
$sel:resourceArn:StopActivityStream'

-> StopActivityStream

Create a value of StopActivityStream with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applyImmediately:StopActivityStream', stopActivityStream_applyImmediately - Specifies whether or not the database activity stream is to stop as soon as possible, regardless of the maintenance window for the database.

$sel:resourceArn:StopActivityStream', stopActivityStream_resourceArn - The Amazon Resource Name (ARN) of the DB cluster for the database activity stream. For example, arn:aws:rds:us-east-1:12345667890:cluster:das-cluster.

Request Lenses

stopActivityStream_applyImmediately :: Lens' StopActivityStream (Maybe Bool)

Specifies whether or not the database activity stream is to stop as soon as possible, regardless of the maintenance window for the database.

stopActivityStream_resourceArn :: Lens' StopActivityStream Text

The Amazon Resource Name (ARN) of the DB cluster for the database activity stream. For example, arn:aws:rds:us-east-1:12345667890:cluster:das-cluster.

Destructuring the Response

data StopActivityStreamResponse

See: newStopActivityStreamResponse smart constructor.

Constructors

StopActivityStreamResponse'

Fields

status :: Maybe ActivityStreamStatus

The status of the database activity stream.

kmsKeyId :: Maybe Text

The Amazon Web Services KMS key identifier used for encrypting messages in the database activity stream.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

kinesisStreamName :: Maybe Text

The name of the Amazon Kinesis data stream used for the database activity stream.

httpStatus :: Int

The response's http status code.

newStopActivityStreamResponse

Arguments

:: Int
$sel:httpStatus:StopActivityStreamResponse'

-> StopActivityStreamResponse

Create a value of StopActivityStreamResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:StopActivityStreamResponse', stopActivityStreamResponse_status - The status of the database activity stream.

$sel:kmsKeyId:StopActivityStreamResponse', stopActivityStreamResponse_kmsKeyId - The Amazon Web Services KMS key identifier used for encrypting messages in the database activity stream.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

$sel:kinesisStreamName:StopActivityStreamResponse', stopActivityStreamResponse_kinesisStreamName - The name of the Amazon Kinesis data stream used for the database activity stream.

$sel:httpStatus:StopActivityStreamResponse', stopActivityStreamResponse_httpStatus - The response's http status code.

Response Lenses

stopActivityStreamResponse_status :: Lens' StopActivityStreamResponse (Maybe ActivityStreamStatus)

The status of the database activity stream.

stopActivityStreamResponse_kmsKeyId :: Lens' StopActivityStreamResponse (Maybe Text)

The Amazon Web Services KMS key identifier used for encrypting messages in the database activity stream.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

stopActivityStreamResponse_kinesisStreamName :: Lens' StopActivityStreamResponse (Maybe Text)

The name of the Amazon Kinesis data stream used for the database activity stream.

stopActivityStreamResponse_httpStatus :: Lens' StopActivityStreamResponse Int

The response's http status code.