Network.AWS.RDS.Types.ConnectionPoolConfiguration

Description

Documentation

data ConnectionPoolConfiguration

Specifies the settings that control the size and behavior of the connection pool associated with a DBProxyTargetGroup.

See: newConnectionPoolConfiguration smart constructor.

Constructors

ConnectionPoolConfiguration'

Fields

maxIdleConnectionsPercent :: Maybe Int

Controls how actively the proxy closes idle database connections in the connection pool. A high value enables the proxy to leave a high percentage of idle connections open. A low value causes the proxy to close idle client connections and return the underlying database connections to the connection pool. For Aurora MySQL, it is expressed as a percentage of the max_connections setting for the RDS DB instance or Aurora DB cluster used by the target group.

Default: 50

Constraints: between 0 and MaxConnectionsPercent

sessionPinningFilters :: Maybe [Text]

Each item in the list represents a class of SQL operations that normally cause all later statements in a session using a proxy to be pinned to the same underlying database connection. Including an item in the list exempts that class of SQL operations from the pinning behavior.

Default: no session pinning filters

maxConnectionsPercent :: Maybe Int

The maximum size of the connection pool for each target in a target group. For Aurora MySQL, it is expressed as a percentage of the max_connections setting for the RDS DB instance or Aurora DB cluster used by the target group.

Default: 100

Constraints: between 1 and 100

connectionBorrowTimeout :: Maybe Int

The number of seconds for a proxy to wait for a connection to become available in the connection pool. Only applies when the proxy has opened its maximum number of connections and all connections are busy with client sessions.

Default: 120

Constraints: between 1 and 3600, or 0 representing unlimited

initQuery :: Maybe Text

One or more SQL statements for the proxy to run when opening each new database connection. Typically used with SET statements to make sure that each connection has identical settings such as time zone and character set. For multiple statements, use semicolons as the separator. You can also include multiple variables in a single SET statement, such as SET x=1, y=2.

Default: no initialization query

newConnectionPoolConfiguration :: ConnectionPoolConfiguration

Create a value of ConnectionPoolConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxIdleConnectionsPercent:ConnectionPoolConfiguration', connectionPoolConfiguration_maxIdleConnectionsPercent - Controls how actively the proxy closes idle database connections in the connection pool. A high value enables the proxy to leave a high percentage of idle connections open. A low value causes the proxy to close idle client connections and return the underlying database connections to the connection pool. For Aurora MySQL, it is expressed as a percentage of the max_connections setting for the RDS DB instance or Aurora DB cluster used by the target group.

Default: 50

Constraints: between 0 and MaxConnectionsPercent

$sel:sessionPinningFilters:ConnectionPoolConfiguration', connectionPoolConfiguration_sessionPinningFilters - Each item in the list represents a class of SQL operations that normally cause all later statements in a session using a proxy to be pinned to the same underlying database connection. Including an item in the list exempts that class of SQL operations from the pinning behavior.

Default: no session pinning filters

$sel:maxConnectionsPercent:ConnectionPoolConfiguration', connectionPoolConfiguration_maxConnectionsPercent - The maximum size of the connection pool for each target in a target group. For Aurora MySQL, it is expressed as a percentage of the max_connections setting for the RDS DB instance or Aurora DB cluster used by the target group.

Default: 100

Constraints: between 1 and 100

$sel:connectionBorrowTimeout:ConnectionPoolConfiguration', connectionPoolConfiguration_connectionBorrowTimeout - The number of seconds for a proxy to wait for a connection to become available in the connection pool. Only applies when the proxy has opened its maximum number of connections and all connections are busy with client sessions.

Default: 120

Constraints: between 1 and 3600, or 0 representing unlimited

$sel:initQuery:ConnectionPoolConfiguration', connectionPoolConfiguration_initQuery - One or more SQL statements for the proxy to run when opening each new database connection. Typically used with SET statements to make sure that each connection has identical settings such as time zone and character set. For multiple statements, use semicolons as the separator. You can also include multiple variables in a single SET statement, such as SET x=1, y=2.

Default: no initialization query

connectionPoolConfiguration_maxIdleConnectionsPercent :: Lens' ConnectionPoolConfiguration (Maybe Int)

Controls how actively the proxy closes idle database connections in the connection pool. A high value enables the proxy to leave a high percentage of idle connections open. A low value causes the proxy to close idle client connections and return the underlying database connections to the connection pool. For Aurora MySQL, it is expressed as a percentage of the max_connections setting for the RDS DB instance or Aurora DB cluster used by the target group.

Default: 50

Constraints: between 0 and MaxConnectionsPercent

connectionPoolConfiguration_sessionPinningFilters :: Lens' ConnectionPoolConfiguration (Maybe [Text])

Each item in the list represents a class of SQL operations that normally cause all later statements in a session using a proxy to be pinned to the same underlying database connection. Including an item in the list exempts that class of SQL operations from the pinning behavior.

Default: no session pinning filters

connectionPoolConfiguration_maxConnectionsPercent :: Lens' ConnectionPoolConfiguration (Maybe Int)

The maximum size of the connection pool for each target in a target group. For Aurora MySQL, it is expressed as a percentage of the max_connections setting for the RDS DB instance or Aurora DB cluster used by the target group.

Default: 100

Constraints: between 1 and 100

connectionPoolConfiguration_connectionBorrowTimeout :: Lens' ConnectionPoolConfiguration (Maybe Int)

The number of seconds for a proxy to wait for a connection to become available in the connection pool. Only applies when the proxy has opened its maximum number of connections and all connections are busy with client sessions.

Default: 120

Constraints: between 1 and 3600, or 0 representing unlimited

connectionPoolConfiguration_initQuery :: Lens' ConnectionPoolConfiguration (Maybe Text)

One or more SQL statements for the proxy to run when opening each new database connection. Typically used with SET statements to make sure that each connection has identical settings such as time zone and character set. For multiple statements, use semicolons as the separator. You can also include multiple variables in a single SET statement, such as SET x=1, y=2.

Default: no initialization query