Network.AWS.RDS.Types.ConnectionPoolConfigurationInfo

Description

Documentation

data ConnectionPoolConfigurationInfo

Displays the settings that control the size and behavior of the connection pool associated with a DBProxyTarget.

See: newConnectionPoolConfigurationInfo smart constructor.

Constructors

ConnectionPoolConfigurationInfo'

Fields

maxIdleConnectionsPercent :: Maybe Int

Controls how actively the proxy closes idle database connections in the connection pool. A high value enables the proxy to leave a high percentage of idle connections open. A low value causes the proxy to close idle client connections and return the underlying database connections to the connection pool. For Aurora MySQL, it is expressed as a percentage of the max_connections setting for the RDS DB instance or Aurora DB cluster used by the target group.

sessionPinningFilters :: Maybe [Text]

Each item in the list represents a class of SQL operations that normally cause all later statements in a session using a proxy to be pinned to the same underlying database connection. Including an item in the list exempts that class of SQL operations from the pinning behavior. Currently, the only allowed value is EXCLUDE_VARIABLE_SETS.

maxConnectionsPercent :: Maybe Int

The maximum size of the connection pool for each target in a target group. For Aurora MySQL, it is expressed as a percentage of the max_connections setting for the RDS DB instance or Aurora DB cluster used by the target group.

connectionBorrowTimeout :: Maybe Int

The number of seconds for a proxy to wait for a connection to become available in the connection pool. Only applies when the proxy has opened its maximum number of connections and all connections are busy with client sessions.

initQuery :: Maybe Text

One or more SQL statements for the proxy to run when opening each new database connection. Typically used with SET statements to make sure that each connection has identical settings such as time zone and character set. This setting is empty by default. For multiple statements, use semicolons as the separator. You can also include multiple variables in a single SET statement, such as SET x=1, y=2.

newConnectionPoolConfigurationInfo :: ConnectionPoolConfigurationInfo

Create a value of ConnectionPoolConfigurationInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxIdleConnectionsPercent:ConnectionPoolConfigurationInfo', connectionPoolConfigurationInfo_maxIdleConnectionsPercent - Controls how actively the proxy closes idle database connections in the connection pool. A high value enables the proxy to leave a high percentage of idle connections open. A low value causes the proxy to close idle client connections and return the underlying database connections to the connection pool. For Aurora MySQL, it is expressed as a percentage of the max_connections setting for the RDS DB instance or Aurora DB cluster used by the target group.

$sel:sessionPinningFilters:ConnectionPoolConfigurationInfo', connectionPoolConfigurationInfo_sessionPinningFilters - Each item in the list represents a class of SQL operations that normally cause all later statements in a session using a proxy to be pinned to the same underlying database connection. Including an item in the list exempts that class of SQL operations from the pinning behavior. Currently, the only allowed value is EXCLUDE_VARIABLE_SETS.

$sel:maxConnectionsPercent:ConnectionPoolConfigurationInfo', connectionPoolConfigurationInfo_maxConnectionsPercent - The maximum size of the connection pool for each target in a target group. For Aurora MySQL, it is expressed as a percentage of the max_connections setting for the RDS DB instance or Aurora DB cluster used by the target group.

$sel:connectionBorrowTimeout:ConnectionPoolConfigurationInfo', connectionPoolConfigurationInfo_connectionBorrowTimeout - The number of seconds for a proxy to wait for a connection to become available in the connection pool. Only applies when the proxy has opened its maximum number of connections and all connections are busy with client sessions.

$sel:initQuery:ConnectionPoolConfigurationInfo', connectionPoolConfigurationInfo_initQuery - One or more SQL statements for the proxy to run when opening each new database connection. Typically used with SET statements to make sure that each connection has identical settings such as time zone and character set. This setting is empty by default. For multiple statements, use semicolons as the separator. You can also include multiple variables in a single SET statement, such as SET x=1, y=2.

connectionPoolConfigurationInfo_maxIdleConnectionsPercent :: Lens' ConnectionPoolConfigurationInfo (Maybe Int)

Controls how actively the proxy closes idle database connections in the connection pool. A high value enables the proxy to leave a high percentage of idle connections open. A low value causes the proxy to close idle client connections and return the underlying database connections to the connection pool. For Aurora MySQL, it is expressed as a percentage of the max_connections setting for the RDS DB instance or Aurora DB cluster used by the target group.

connectionPoolConfigurationInfo_sessionPinningFilters :: Lens' ConnectionPoolConfigurationInfo (Maybe [Text])

Each item in the list represents a class of SQL operations that normally cause all later statements in a session using a proxy to be pinned to the same underlying database connection. Including an item in the list exempts that class of SQL operations from the pinning behavior. Currently, the only allowed value is EXCLUDE_VARIABLE_SETS.

connectionPoolConfigurationInfo_maxConnectionsPercent :: Lens' ConnectionPoolConfigurationInfo (Maybe Int)

The maximum size of the connection pool for each target in a target group. For Aurora MySQL, it is expressed as a percentage of the max_connections setting for the RDS DB instance or Aurora DB cluster used by the target group.

connectionPoolConfigurationInfo_connectionBorrowTimeout :: Lens' ConnectionPoolConfigurationInfo (Maybe Int)

The number of seconds for a proxy to wait for a connection to become available in the connection pool. Only applies when the proxy has opened its maximum number of connections and all connections are busy with client sessions.

connectionPoolConfigurationInfo_initQuery :: Lens' ConnectionPoolConfigurationInfo (Maybe Text)

One or more SQL statements for the proxy to run when opening each new database connection. Typically used with SET statements to make sure that each connection has identical settings such as time zone and character set. This setting is empty by default. For multiple statements, use semicolons as the separator. You can also include multiple variables in a single SET statement, such as SET x=1, y=2.