Network.AWS.RDS.Types.CustomAvailabilityZone

Description

Documentation

data CustomAvailabilityZone

A custom Availability Zone (AZ) is an on-premises AZ that is integrated with a VMware vSphere cluster.

For more information about RDS on VMware, see the RDS on VMware User Guide.

See: newCustomAvailabilityZone smart constructor.

Constructors

CustomAvailabilityZone'

Fields

vpnDetails :: Maybe VpnDetails

Information about the virtual private network (VPN) between the VMware vSphere cluster and the Amazon Web Services website.

customAvailabilityZoneName :: Maybe Text

The name of the custom AZ.

customAvailabilityZoneId :: Maybe Text

The identifier of the custom AZ.

Amazon RDS generates a unique identifier when a custom AZ is created.

customAvailabilityZoneStatus :: Maybe Text

The status of the custom AZ.

newCustomAvailabilityZone :: CustomAvailabilityZone

Create a value of CustomAvailabilityZone with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpnDetails:CustomAvailabilityZone', customAvailabilityZone_vpnDetails - Information about the virtual private network (VPN) between the VMware vSphere cluster and the Amazon Web Services website.

$sel:customAvailabilityZoneName:CustomAvailabilityZone', customAvailabilityZone_customAvailabilityZoneName - The name of the custom AZ.

$sel:customAvailabilityZoneId:CustomAvailabilityZone', customAvailabilityZone_customAvailabilityZoneId - The identifier of the custom AZ.

Amazon RDS generates a unique identifier when a custom AZ is created.

$sel:customAvailabilityZoneStatus:CustomAvailabilityZone', customAvailabilityZone_customAvailabilityZoneStatus - The status of the custom AZ.

customAvailabilityZone_vpnDetails :: Lens' CustomAvailabilityZone (Maybe VpnDetails)

Information about the virtual private network (VPN) between the VMware vSphere cluster and the Amazon Web Services website.

customAvailabilityZone_customAvailabilityZoneName :: Lens' CustomAvailabilityZone (Maybe Text)

The name of the custom AZ.

customAvailabilityZone_customAvailabilityZoneId :: Lens' CustomAvailabilityZone (Maybe Text)

The identifier of the custom AZ.

Amazon RDS generates a unique identifier when a custom AZ is created.

customAvailabilityZone_customAvailabilityZoneStatus :: Lens' CustomAvailabilityZone (Maybe Text)

The status of the custom AZ.