Network.AWS.RDS.Types.DBCluster

Description

Documentation

data DBCluster

Contains the details of an Amazon Aurora DB cluster.

This data type is used as a response element in the DescribeDBClusters, StopDBCluster, and StartDBCluster actions.

See: newDBCluster smart constructor.

Constructors

DBCluster'

Fields

associatedRoles :: Maybe [DBClusterRole]

Provides a list of the Amazon Web Services Identity and Access Management (IAM) roles that are associated with the DB cluster. IAM roles that are associated with a DB cluster grant permission for the DB cluster to access other Amazon Web Services on your behalf.

vpcSecurityGroups :: Maybe [VpcSecurityGroupMembership]

Provides a list of VPC security groups that the DB cluster belongs to.

automaticRestartTime :: Maybe ISO8601

The time when a stopped DB cluster is restarted automatically.

deletionProtection :: Maybe Bool

Indicates if the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled.

preferredBackupWindow :: Maybe Text

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

dbClusterMembers :: Maybe [DBClusterMember]

Provides the list of instances that make up the DB cluster.

backupRetentionPeriod :: Maybe Int

Specifies the number of days for which automatic DB snapshots are retained.

storageEncrypted :: Maybe Bool

Specifies whether the DB cluster is encrypted.

availabilityZones :: Maybe [Text]

Provides the list of Availability Zones (AZs) where instances in the DB cluster can be created.

status :: Maybe Text

Specifies the current state of this DB cluster.

backtrackConsumedChangeRecords :: Maybe Integer

The number of change records stored for Backtrack.

crossAccountClone :: Maybe Bool

Specifies whether the DB cluster is a clone of a DB cluster owned by a different Amazon Web Services account.

globalWriteForwardingStatus :: Maybe WriteForwardingStatus

Specifies whether a secondary cluster in an Aurora global database has write forwarding enabled, not enabled, or is in the process of enabling it.

domainMemberships :: Maybe [DomainMembership]

The Active Directory Domain membership records associated with the DB cluster.

dbClusterArn :: Maybe Text

The Amazon Resource Name (ARN) for the DB cluster.

cloneGroupId :: Maybe Text

Identifies the clone group to which the DB cluster is associated.

dbClusterOptionGroupMemberships :: Maybe [DBClusterOptionGroupStatus]

Provides the list of option group memberships for this DB cluster.

latestRestorableTime :: Maybe ISO8601

Specifies the latest time to which a database can be restored with point-in-time restore.

readerEndpoint :: Maybe Text

The reader endpoint for the DB cluster. The reader endpoint for a DB cluster load-balances connections across the Aurora Replicas that are available in a DB cluster. As clients request new connections to the reader endpoint, Aurora distributes the connection requests among the Aurora Replicas in the DB cluster. This functionality can help balance your read workload across multiple Aurora Replicas in your DB cluster.

If a failover occurs, and the Aurora Replica that you are connected to is promoted to be the primary instance, your connection is dropped. To continue sending your read workload to other Aurora Replicas in the cluster, you can then reconnect to the reader endpoint.

activityStreamKmsKeyId :: Maybe Text

The Amazon Web Services KMS key identifier used for encrypting messages in the database activity stream.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

httpEndpointEnabled :: Maybe Bool

A value that indicates whether the HTTP endpoint for an Aurora Serverless DB cluster is enabled.

When enabled, the HTTP endpoint provides a connectionless web service API for running SQL queries on the Aurora Serverless DB cluster. You can also query your database from inside the RDS console with the query editor.

For more information, see Using the Data API for Aurora Serverless in the Amazon Aurora User Guide.

clusterCreateTime :: Maybe ISO8601

Specifies the time when the DB cluster was created, in Universal Coordinated Time (UTC).

earliestRestorableTime :: Maybe ISO8601

The earliest time to which a database can be restored with point-in-time restore.

enabledCloudwatchLogsExports :: Maybe [Text]

A list of log types that this DB cluster is configured to export to CloudWatch Logs.

Log types vary by DB engine. For information about the log types for each DB engine, see Amazon RDS Database Log Files in the Amazon Aurora User Guide.

activityStreamStatus :: Maybe ActivityStreamStatus

The status of the database activity stream.

engineMode :: Maybe Text

The DB engine mode of the DB cluster, either provisioned, serverless, parallelquery, global, or multimaster.

For more information, see CreateDBCluster.

multiAZ :: Maybe Bool

Specifies whether the DB cluster has instances in multiple Availability Zones.

dbSubnetGroup :: Maybe Text

Specifies information on the subnet group associated with the DB cluster, including the name, description, and subnets in the subnet group.

readReplicaIdentifiers :: Maybe [Text]

Contains one or more identifiers of the read replicas associated with this DB cluster.

masterUsername :: Maybe Text

Contains the master username for the DB cluster.

hostedZoneId :: Maybe Text

Specifies the ID that Amazon Route 53 assigns when you create a hosted zone.

kmsKeyId :: Maybe Text

If StorageEncrypted is enabled, the Amazon Web Services KMS key identifier for the encrypted DB cluster.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

dbClusterIdentifier :: Maybe Text

Contains a user-supplied DB cluster identifier. This identifier is the unique key that identifies a DB cluster.

engineVersion :: Maybe Text

Indicates the database engine version.

capacity :: Maybe Int

The current capacity of an Aurora Serverless DB cluster. The capacity is 0 (zero) when the cluster is paused.

For more information about Aurora Serverless, see Using Amazon Aurora Serverless in the Amazon Aurora User Guide.

characterSetName :: Maybe Text

If present, specifies the name of the character set that this cluster is associated with.

preferredMaintenanceWindow :: Maybe Text

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

port :: Maybe Int

Specifies the port that the database engine is listening on.

percentProgress :: Maybe Text

Specifies the progress of the operation as a percentage.

engine :: Maybe Text

The name of the database engine to be used for this DB cluster.

pendingModifiedValues :: Maybe ClusterPendingModifiedValues

A value that specifies that changes to the DB cluster are pending. This element is only included when changes are pending. Specific changes are identified by subelements.

endpoint :: Maybe Text

Specifies the connection endpoint for the primary instance of the DB cluster.

copyTagsToSnapshot :: Maybe Bool

Specifies whether tags are copied from the DB cluster to snapshots of the DB cluster.

scalingConfigurationInfo :: Maybe ScalingConfigurationInfo

dbClusterResourceId :: Maybe Text

The Amazon Web Services Region-unique, immutable identifier for the DB cluster. This identifier is found in Amazon Web Services CloudTrail log entries whenever the Amazon Web Services KMS CMK for the DB cluster is accessed.

customEndpoints :: Maybe [Text]

Identifies all custom endpoints associated with the cluster.

tagList :: Maybe [Tag]

earliestBacktrackTime :: Maybe ISO8601

The earliest time to which a DB cluster can be backtracked.

backtrackWindow :: Maybe Integer

The target backtrack window, in seconds. If this value is set to 0, backtracking is disabled for the DB cluster. Otherwise, backtracking is enabled.

allocatedStorage :: Maybe Int

For all database engines except Amazon Aurora, AllocatedStorage specifies the allocated storage size in gibibytes (GiB). For Aurora, AllocatedStorage always returns 1, because Aurora DB cluster storage size isn't fixed, but instead automatically adjusts as needed.

globalWriteForwardingRequested :: Maybe Bool

Specifies whether you have requested to enable write forwarding for a secondary cluster in an Aurora global database. Because write forwarding takes time to enable, check the value of GlobalWriteForwardingStatus to confirm that the request has completed before using the write forwarding feature for this cluster.

iAMDatabaseAuthenticationEnabled :: Maybe Bool

A value that indicates whether the mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts is enabled.

dbClusterParameterGroup :: Maybe Text

Specifies the name of the DB cluster parameter group for the DB cluster.

replicationSourceIdentifier :: Maybe Text

Contains the identifier of the source DB cluster if this DB cluster is a read replica.

activityStreamKinesisStreamName :: Maybe Text

The name of the Amazon Kinesis data stream used for the database activity stream.

databaseName :: Maybe Text

Contains the name of the initial database of this DB cluster that was provided at create time, if one was specified when the DB cluster was created. This same name is returned for the life of the DB cluster.

activityStreamMode :: Maybe ActivityStreamMode

The mode of the database activity stream. Database events such as a change or access generate an activity stream event. The database session can handle these events either synchronously or asynchronously.

newDBCluster :: DBCluster

Create a value of DBCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associatedRoles:DBCluster', dbCluster_associatedRoles - Provides a list of the Amazon Web Services Identity and Access Management (IAM) roles that are associated with the DB cluster. IAM roles that are associated with a DB cluster grant permission for the DB cluster to access other Amazon Web Services on your behalf.

$sel:vpcSecurityGroups:DBCluster', dbCluster_vpcSecurityGroups - Provides a list of VPC security groups that the DB cluster belongs to.

$sel:automaticRestartTime:DBCluster', dbCluster_automaticRestartTime - The time when a stopped DB cluster is restarted automatically.

$sel:deletionProtection:DBCluster', dbCluster_deletionProtection - Indicates if the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled.

$sel:preferredBackupWindow:DBCluster', dbCluster_preferredBackupWindow - Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

$sel:dbClusterMembers:DBCluster', dbCluster_dbClusterMembers - Provides the list of instances that make up the DB cluster.

$sel:backupRetentionPeriod:DBCluster', dbCluster_backupRetentionPeriod - Specifies the number of days for which automatic DB snapshots are retained.

$sel:storageEncrypted:DBCluster', dbCluster_storageEncrypted - Specifies whether the DB cluster is encrypted.

$sel:availabilityZones:DBCluster', dbCluster_availabilityZones - Provides the list of Availability Zones (AZs) where instances in the DB cluster can be created.

$sel:status:DBCluster', dbCluster_status - Specifies the current state of this DB cluster.

$sel:backtrackConsumedChangeRecords:DBCluster', dbCluster_backtrackConsumedChangeRecords - The number of change records stored for Backtrack.

$sel:crossAccountClone:DBCluster', dbCluster_crossAccountClone - Specifies whether the DB cluster is a clone of a DB cluster owned by a different Amazon Web Services account.

$sel:globalWriteForwardingStatus:DBCluster', dbCluster_globalWriteForwardingStatus - Specifies whether a secondary cluster in an Aurora global database has write forwarding enabled, not enabled, or is in the process of enabling it.

$sel:domainMemberships:DBCluster', dbCluster_domainMemberships - The Active Directory Domain membership records associated with the DB cluster.

$sel:dbClusterArn:DBCluster', dbCluster_dbClusterArn - The Amazon Resource Name (ARN) for the DB cluster.

$sel:cloneGroupId:DBCluster', dbCluster_cloneGroupId - Identifies the clone group to which the DB cluster is associated.

$sel:dbClusterOptionGroupMemberships:DBCluster', dbCluster_dbClusterOptionGroupMemberships - Provides the list of option group memberships for this DB cluster.

$sel:latestRestorableTime:DBCluster', dbCluster_latestRestorableTime - Specifies the latest time to which a database can be restored with point-in-time restore.

$sel:readerEndpoint:DBCluster', dbCluster_readerEndpoint - The reader endpoint for the DB cluster. The reader endpoint for a DB cluster load-balances connections across the Aurora Replicas that are available in a DB cluster. As clients request new connections to the reader endpoint, Aurora distributes the connection requests among the Aurora Replicas in the DB cluster. This functionality can help balance your read workload across multiple Aurora Replicas in your DB cluster.

If a failover occurs, and the Aurora Replica that you are connected to is promoted to be the primary instance, your connection is dropped. To continue sending your read workload to other Aurora Replicas in the cluster, you can then reconnect to the reader endpoint.

$sel:activityStreamKmsKeyId:DBCluster', dbCluster_activityStreamKmsKeyId - The Amazon Web Services KMS key identifier used for encrypting messages in the database activity stream.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

$sel:httpEndpointEnabled:DBCluster', dbCluster_httpEndpointEnabled - A value that indicates whether the HTTP endpoint for an Aurora Serverless DB cluster is enabled.

When enabled, the HTTP endpoint provides a connectionless web service API for running SQL queries on the Aurora Serverless DB cluster. You can also query your database from inside the RDS console with the query editor.

For more information, see Using the Data API for Aurora Serverless in the Amazon Aurora User Guide.

$sel:clusterCreateTime:DBCluster', dbCluster_clusterCreateTime - Specifies the time when the DB cluster was created, in Universal Coordinated Time (UTC).

$sel:earliestRestorableTime:DBCluster', dbCluster_earliestRestorableTime - The earliest time to which a database can be restored with point-in-time restore.

$sel:enabledCloudwatchLogsExports:DBCluster', dbCluster_enabledCloudwatchLogsExports - A list of log types that this DB cluster is configured to export to CloudWatch Logs.

Log types vary by DB engine. For information about the log types for each DB engine, see Amazon RDS Database Log Files in the Amazon Aurora User Guide.

$sel:activityStreamStatus:DBCluster', dbCluster_activityStreamStatus - The status of the database activity stream.

$sel:engineMode:DBCluster', dbCluster_engineMode - The DB engine mode of the DB cluster, either provisioned, serverless, parallelquery, global, or multimaster.

For more information, see CreateDBCluster.

$sel:multiAZ:DBCluster', dbCluster_multiAZ - Specifies whether the DB cluster has instances in multiple Availability Zones.

$sel:dbSubnetGroup:DBCluster', dbCluster_dbSubnetGroup - Specifies information on the subnet group associated with the DB cluster, including the name, description, and subnets in the subnet group.

$sel:readReplicaIdentifiers:DBCluster', dbCluster_readReplicaIdentifiers - Contains one or more identifiers of the read replicas associated with this DB cluster.

$sel:masterUsername:DBCluster', dbCluster_masterUsername - Contains the master username for the DB cluster.

$sel:hostedZoneId:DBCluster', dbCluster_hostedZoneId - Specifies the ID that Amazon Route 53 assigns when you create a hosted zone.

$sel:kmsKeyId:DBCluster', dbCluster_kmsKeyId - If StorageEncrypted is enabled, the Amazon Web Services KMS key identifier for the encrypted DB cluster.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

$sel:dbClusterIdentifier:DBCluster', dbCluster_dbClusterIdentifier - Contains a user-supplied DB cluster identifier. This identifier is the unique key that identifies a DB cluster.

$sel:engineVersion:DBCluster', dbCluster_engineVersion - Indicates the database engine version.

$sel:capacity:DBCluster', dbCluster_capacity - The current capacity of an Aurora Serverless DB cluster. The capacity is 0 (zero) when the cluster is paused.

For more information about Aurora Serverless, see Using Amazon Aurora Serverless in the Amazon Aurora User Guide.

$sel:characterSetName:DBCluster', dbCluster_characterSetName - If present, specifies the name of the character set that this cluster is associated with.

$sel:preferredMaintenanceWindow:DBCluster', dbCluster_preferredMaintenanceWindow - Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

$sel:port:DBCluster', dbCluster_port - Specifies the port that the database engine is listening on.

$sel:percentProgress:DBCluster', dbCluster_percentProgress - Specifies the progress of the operation as a percentage.

$sel:engine:DBCluster', dbCluster_engine - The name of the database engine to be used for this DB cluster.

$sel:pendingModifiedValues:DBCluster', dbCluster_pendingModifiedValues - A value that specifies that changes to the DB cluster are pending. This element is only included when changes are pending. Specific changes are identified by subelements.

$sel:endpoint:DBCluster', dbCluster_endpoint - Specifies the connection endpoint for the primary instance of the DB cluster.

$sel:copyTagsToSnapshot:DBCluster', dbCluster_copyTagsToSnapshot - Specifies whether tags are copied from the DB cluster to snapshots of the DB cluster.

$sel:scalingConfigurationInfo:DBCluster', dbCluster_scalingConfigurationInfo - Undocumented member.

$sel:dbClusterResourceId:DBCluster', dbCluster_dbClusterResourceId - The Amazon Web Services Region-unique, immutable identifier for the DB cluster. This identifier is found in Amazon Web Services CloudTrail log entries whenever the Amazon Web Services KMS CMK for the DB cluster is accessed.

$sel:customEndpoints:DBCluster', dbCluster_customEndpoints - Identifies all custom endpoints associated with the cluster.

$sel:tagList:DBCluster', dbCluster_tagList - Undocumented member.

$sel:earliestBacktrackTime:DBCluster', dbCluster_earliestBacktrackTime - The earliest time to which a DB cluster can be backtracked.

$sel:backtrackWindow:DBCluster', dbCluster_backtrackWindow - The target backtrack window, in seconds. If this value is set to 0, backtracking is disabled for the DB cluster. Otherwise, backtracking is enabled.

$sel:allocatedStorage:DBCluster', dbCluster_allocatedStorage - For all database engines except Amazon Aurora, AllocatedStorage specifies the allocated storage size in gibibytes (GiB). For Aurora, AllocatedStorage always returns 1, because Aurora DB cluster storage size isn't fixed, but instead automatically adjusts as needed.

$sel:globalWriteForwardingRequested:DBCluster', dbCluster_globalWriteForwardingRequested - Specifies whether you have requested to enable write forwarding for a secondary cluster in an Aurora global database. Because write forwarding takes time to enable, check the value of GlobalWriteForwardingStatus to confirm that the request has completed before using the write forwarding feature for this cluster.

$sel:iAMDatabaseAuthenticationEnabled:DBCluster', dbCluster_iAMDatabaseAuthenticationEnabled - A value that indicates whether the mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts is enabled.

$sel:dbClusterParameterGroup:DBCluster', dbCluster_dbClusterParameterGroup - Specifies the name of the DB cluster parameter group for the DB cluster.

$sel:replicationSourceIdentifier:DBCluster', dbCluster_replicationSourceIdentifier - Contains the identifier of the source DB cluster if this DB cluster is a read replica.

$sel:activityStreamKinesisStreamName:DBCluster', dbCluster_activityStreamKinesisStreamName - The name of the Amazon Kinesis data stream used for the database activity stream.

$sel:databaseName:DBCluster', dbCluster_databaseName - Contains the name of the initial database of this DB cluster that was provided at create time, if one was specified when the DB cluster was created. This same name is returned for the life of the DB cluster.

$sel:activityStreamMode:DBCluster', dbCluster_activityStreamMode - The mode of the database activity stream. Database events such as a change or access generate an activity stream event. The database session can handle these events either synchronously or asynchronously.

dbCluster_associatedRoles :: Lens' DBCluster (Maybe [DBClusterRole])

Provides a list of the Amazon Web Services Identity and Access Management (IAM) roles that are associated with the DB cluster. IAM roles that are associated with a DB cluster grant permission for the DB cluster to access other Amazon Web Services on your behalf.

dbCluster_vpcSecurityGroups :: Lens' DBCluster (Maybe [VpcSecurityGroupMembership])

Provides a list of VPC security groups that the DB cluster belongs to.

dbCluster_automaticRestartTime :: Lens' DBCluster (Maybe UTCTime)

The time when a stopped DB cluster is restarted automatically.

dbCluster_deletionProtection :: Lens' DBCluster (Maybe Bool)

Indicates if the DB cluster has deletion protection enabled. The database can't be deleted when deletion protection is enabled.

dbCluster_preferredBackupWindow :: Lens' DBCluster (Maybe Text)

Specifies the daily time range during which automated backups are created if automated backups are enabled, as determined by the BackupRetentionPeriod.

dbCluster_dbClusterMembers :: Lens' DBCluster (Maybe [DBClusterMember])

Provides the list of instances that make up the DB cluster.

dbCluster_backupRetentionPeriod :: Lens' DBCluster (Maybe Int)

Specifies the number of days for which automatic DB snapshots are retained.

dbCluster_storageEncrypted :: Lens' DBCluster (Maybe Bool)

Specifies whether the DB cluster is encrypted.

dbCluster_availabilityZones :: Lens' DBCluster (Maybe [Text])

Provides the list of Availability Zones (AZs) where instances in the DB cluster can be created.

dbCluster_status :: Lens' DBCluster (Maybe Text)

Specifies the current state of this DB cluster.

dbCluster_backtrackConsumedChangeRecords :: Lens' DBCluster (Maybe Integer)

The number of change records stored for Backtrack.

dbCluster_crossAccountClone :: Lens' DBCluster (Maybe Bool)

Specifies whether the DB cluster is a clone of a DB cluster owned by a different Amazon Web Services account.

dbCluster_globalWriteForwardingStatus :: Lens' DBCluster (Maybe WriteForwardingStatus)

Specifies whether a secondary cluster in an Aurora global database has write forwarding enabled, not enabled, or is in the process of enabling it.

dbCluster_domainMemberships :: Lens' DBCluster (Maybe [DomainMembership])

The Active Directory Domain membership records associated with the DB cluster.

dbCluster_dbClusterArn :: Lens' DBCluster (Maybe Text)

The Amazon Resource Name (ARN) for the DB cluster.

dbCluster_cloneGroupId :: Lens' DBCluster (Maybe Text)

Identifies the clone group to which the DB cluster is associated.

dbCluster_dbClusterOptionGroupMemberships :: Lens' DBCluster (Maybe [DBClusterOptionGroupStatus])

Provides the list of option group memberships for this DB cluster.

dbCluster_latestRestorableTime :: Lens' DBCluster (Maybe UTCTime)

Specifies the latest time to which a database can be restored with point-in-time restore.

dbCluster_readerEndpoint :: Lens' DBCluster (Maybe Text)

The reader endpoint for the DB cluster. The reader endpoint for a DB cluster load-balances connections across the Aurora Replicas that are available in a DB cluster. As clients request new connections to the reader endpoint, Aurora distributes the connection requests among the Aurora Replicas in the DB cluster. This functionality can help balance your read workload across multiple Aurora Replicas in your DB cluster.

If a failover occurs, and the Aurora Replica that you are connected to is promoted to be the primary instance, your connection is dropped. To continue sending your read workload to other Aurora Replicas in the cluster, you can then reconnect to the reader endpoint.

dbCluster_activityStreamKmsKeyId :: Lens' DBCluster (Maybe Text)

The Amazon Web Services KMS key identifier used for encrypting messages in the database activity stream.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

dbCluster_httpEndpointEnabled :: Lens' DBCluster (Maybe Bool)

A value that indicates whether the HTTP endpoint for an Aurora Serverless DB cluster is enabled.

When enabled, the HTTP endpoint provides a connectionless web service API for running SQL queries on the Aurora Serverless DB cluster. You can also query your database from inside the RDS console with the query editor.

For more information, see Using the Data API for Aurora Serverless in the Amazon Aurora User Guide.

dbCluster_clusterCreateTime :: Lens' DBCluster (Maybe UTCTime)

Specifies the time when the DB cluster was created, in Universal Coordinated Time (UTC).

dbCluster_earliestRestorableTime :: Lens' DBCluster (Maybe UTCTime)

The earliest time to which a database can be restored with point-in-time restore.

dbCluster_enabledCloudwatchLogsExports :: Lens' DBCluster (Maybe [Text])

A list of log types that this DB cluster is configured to export to CloudWatch Logs.

Log types vary by DB engine. For information about the log types for each DB engine, see Amazon RDS Database Log Files in the Amazon Aurora User Guide.

dbCluster_activityStreamStatus :: Lens' DBCluster (Maybe ActivityStreamStatus)

The status of the database activity stream.

dbCluster_engineMode :: Lens' DBCluster (Maybe Text)

The DB engine mode of the DB cluster, either provisioned, serverless, parallelquery, global, or multimaster.

For more information, see CreateDBCluster.

dbCluster_multiAZ :: Lens' DBCluster (Maybe Bool)

Specifies whether the DB cluster has instances in multiple Availability Zones.

dbCluster_dbSubnetGroup :: Lens' DBCluster (Maybe Text)

Specifies information on the subnet group associated with the DB cluster, including the name, description, and subnets in the subnet group.

dbCluster_readReplicaIdentifiers :: Lens' DBCluster (Maybe [Text])

Contains one or more identifiers of the read replicas associated with this DB cluster.

dbCluster_masterUsername :: Lens' DBCluster (Maybe Text)

Contains the master username for the DB cluster.

dbCluster_hostedZoneId :: Lens' DBCluster (Maybe Text)

Specifies the ID that Amazon Route 53 assigns when you create a hosted zone.

dbCluster_kmsKeyId :: Lens' DBCluster (Maybe Text)

If StorageEncrypted is enabled, the Amazon Web Services KMS key identifier for the encrypted DB cluster.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

dbCluster_dbClusterIdentifier :: Lens' DBCluster (Maybe Text)

Contains a user-supplied DB cluster identifier. This identifier is the unique key that identifies a DB cluster.

dbCluster_engineVersion :: Lens' DBCluster (Maybe Text)

Indicates the database engine version.

dbCluster_capacity :: Lens' DBCluster (Maybe Int)

The current capacity of an Aurora Serverless DB cluster. The capacity is 0 (zero) when the cluster is paused.

For more information about Aurora Serverless, see Using Amazon Aurora Serverless in the Amazon Aurora User Guide.

dbCluster_characterSetName :: Lens' DBCluster (Maybe Text)

If present, specifies the name of the character set that this cluster is associated with.

dbCluster_preferredMaintenanceWindow :: Lens' DBCluster (Maybe Text)

Specifies the weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC).

dbCluster_port :: Lens' DBCluster (Maybe Int)

Specifies the port that the database engine is listening on.

dbCluster_percentProgress :: Lens' DBCluster (Maybe Text)

Specifies the progress of the operation as a percentage.

dbCluster_engine :: Lens' DBCluster (Maybe Text)

The name of the database engine to be used for this DB cluster.

dbCluster_pendingModifiedValues :: Lens' DBCluster (Maybe ClusterPendingModifiedValues)

A value that specifies that changes to the DB cluster are pending. This element is only included when changes are pending. Specific changes are identified by subelements.

dbCluster_endpoint :: Lens' DBCluster (Maybe Text)

Specifies the connection endpoint for the primary instance of the DB cluster.

dbCluster_copyTagsToSnapshot :: Lens' DBCluster (Maybe Bool)

Specifies whether tags are copied from the DB cluster to snapshots of the DB cluster.

dbCluster_scalingConfigurationInfo :: Lens' DBCluster (Maybe ScalingConfigurationInfo)

Undocumented member.

dbCluster_dbClusterResourceId :: Lens' DBCluster (Maybe Text)

The Amazon Web Services Region-unique, immutable identifier for the DB cluster. This identifier is found in Amazon Web Services CloudTrail log entries whenever the Amazon Web Services KMS CMK for the DB cluster is accessed.

dbCluster_customEndpoints :: Lens' DBCluster (Maybe [Text])

Identifies all custom endpoints associated with the cluster.

dbCluster_tagList :: Lens' DBCluster (Maybe [Tag])

Undocumented member.

dbCluster_earliestBacktrackTime :: Lens' DBCluster (Maybe UTCTime)

The earliest time to which a DB cluster can be backtracked.

dbCluster_backtrackWindow :: Lens' DBCluster (Maybe Integer)

The target backtrack window, in seconds. If this value is set to 0, backtracking is disabled for the DB cluster. Otherwise, backtracking is enabled.

dbCluster_allocatedStorage :: Lens' DBCluster (Maybe Int)

For all database engines except Amazon Aurora, AllocatedStorage specifies the allocated storage size in gibibytes (GiB). For Aurora, AllocatedStorage always returns 1, because Aurora DB cluster storage size isn't fixed, but instead automatically adjusts as needed.

dbCluster_globalWriteForwardingRequested :: Lens' DBCluster (Maybe Bool)

Specifies whether you have requested to enable write forwarding for a secondary cluster in an Aurora global database. Because write forwarding takes time to enable, check the value of GlobalWriteForwardingStatus to confirm that the request has completed before using the write forwarding feature for this cluster.

dbCluster_iAMDatabaseAuthenticationEnabled :: Lens' DBCluster (Maybe Bool)

A value that indicates whether the mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts is enabled.

dbCluster_dbClusterParameterGroup :: Lens' DBCluster (Maybe Text)

Specifies the name of the DB cluster parameter group for the DB cluster.

dbCluster_replicationSourceIdentifier :: Lens' DBCluster (Maybe Text)

Contains the identifier of the source DB cluster if this DB cluster is a read replica.

dbCluster_activityStreamKinesisStreamName :: Lens' DBCluster (Maybe Text)

The name of the Amazon Kinesis data stream used for the database activity stream.

dbCluster_databaseName :: Lens' DBCluster (Maybe Text)

Contains the name of the initial database of this DB cluster that was provided at create time, if one was specified when the DB cluster was created. This same name is returned for the life of the DB cluster.

dbCluster_activityStreamMode :: Lens' DBCluster (Maybe ActivityStreamMode)

The mode of the database activity stream. Database events such as a change or access generate an activity stream event. The database session can handle these events either synchronously or asynchronously.