Network.AWS.RDS.Types.DBClusterBacktrack

Description

Documentation

data DBClusterBacktrack

This data type is used as a response element in the DescribeDBClusterBacktracks action.

See: newDBClusterBacktrack smart constructor.

Constructors

DBClusterBacktrack'

Fields

status :: Maybe Text

The status of the backtrack. This property returns one of the following values:

applying - The backtrack is currently being applied to or rolled back from the DB cluster.

completed - The backtrack has successfully been applied to or rolled back from the DB cluster.

failed - An error occurred while the backtrack was applied to or rolled back from the DB cluster.

pending - The backtrack is currently pending application to or rollback from the DB cluster.

backtrackTo :: Maybe ISO8601

The timestamp of the time to which the DB cluster was backtracked.

dbClusterIdentifier :: Maybe Text

Contains a user-supplied DB cluster identifier. This identifier is the unique key that identifies a DB cluster.

backtrackIdentifier :: Maybe Text

Contains the backtrack identifier.

backtrackRequestCreationTime :: Maybe ISO8601

The timestamp of the time at which the backtrack was requested.

backtrackedFrom :: Maybe ISO8601

The timestamp of the time from which the DB cluster was backtracked.

newDBClusterBacktrack :: DBClusterBacktrack

Create a value of DBClusterBacktrack with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DBClusterBacktrack', dbClusterBacktrack_status - The status of the backtrack. This property returns one of the following values:

applying - The backtrack is currently being applied to or rolled back from the DB cluster.

completed - The backtrack has successfully been applied to or rolled back from the DB cluster.

failed - An error occurred while the backtrack was applied to or rolled back from the DB cluster.

pending - The backtrack is currently pending application to or rollback from the DB cluster.

$sel:backtrackTo:DBClusterBacktrack', dbClusterBacktrack_backtrackTo - The timestamp of the time to which the DB cluster was backtracked.

$sel:dbClusterIdentifier:DBClusterBacktrack', dbClusterBacktrack_dbClusterIdentifier - Contains a user-supplied DB cluster identifier. This identifier is the unique key that identifies a DB cluster.

$sel:backtrackIdentifier:DBClusterBacktrack', dbClusterBacktrack_backtrackIdentifier - Contains the backtrack identifier.

$sel:backtrackRequestCreationTime:DBClusterBacktrack', dbClusterBacktrack_backtrackRequestCreationTime - The timestamp of the time at which the backtrack was requested.

$sel:backtrackedFrom:DBClusterBacktrack', dbClusterBacktrack_backtrackedFrom - The timestamp of the time from which the DB cluster was backtracked.

dbClusterBacktrack_status :: Lens' DBClusterBacktrack (Maybe Text)

The status of the backtrack. This property returns one of the following values:

applying - The backtrack is currently being applied to or rolled back from the DB cluster.

completed - The backtrack has successfully been applied to or rolled back from the DB cluster.

failed - An error occurred while the backtrack was applied to or rolled back from the DB cluster.

pending - The backtrack is currently pending application to or rollback from the DB cluster.

dbClusterBacktrack_backtrackTo :: Lens' DBClusterBacktrack (Maybe UTCTime)

The timestamp of the time to which the DB cluster was backtracked.

dbClusterBacktrack_dbClusterIdentifier :: Lens' DBClusterBacktrack (Maybe Text)

Contains a user-supplied DB cluster identifier. This identifier is the unique key that identifies a DB cluster.

dbClusterBacktrack_backtrackIdentifier :: Lens' DBClusterBacktrack (Maybe Text)

Contains the backtrack identifier.

dbClusterBacktrack_backtrackRequestCreationTime :: Lens' DBClusterBacktrack (Maybe UTCTime)

The timestamp of the time at which the backtrack was requested.

dbClusterBacktrack_backtrackedFrom :: Lens' DBClusterBacktrack (Maybe UTCTime)

The timestamp of the time from which the DB cluster was backtracked.