Network.AWS.RDS.Types.DBClusterEndpoint

Description

Documentation

data DBClusterEndpoint

This data type represents the information you need to connect to an Amazon Aurora DB cluster. This data type is used as a response element in the following actions:

CreateDBClusterEndpoint

DescribeDBClusterEndpoints

ModifyDBClusterEndpoint

DeleteDBClusterEndpoint

For the data structure that represents Amazon RDS DB instance endpoints, see Endpoint.

See: newDBClusterEndpoint smart constructor.

Constructors

DBClusterEndpoint'

Fields

status :: Maybe Text

The current status of the endpoint. One of: creating, available, deleting, inactive, modifying. The inactive state applies to an endpoint that can't be used for a certain kind of cluster, such as a writer endpoint for a read-only secondary cluster in a global database.

dbClusterEndpointIdentifier :: Maybe Text

The identifier associated with the endpoint. This parameter is stored as a lowercase string.

excludedMembers :: Maybe [Text]

List of DB instance identifiers that aren't part of the custom endpoint group. All other eligible instances are reachable through the custom endpoint. Only relevant if the list of static members is empty.

endpointType :: Maybe Text

The type of the endpoint. One of: READER, WRITER, CUSTOM.

customEndpointType :: Maybe Text

The type associated with a custom endpoint. One of: READER, WRITER, ANY.

dbClusterEndpointArn :: Maybe Text

The Amazon Resource Name (ARN) for the endpoint.

dbClusterIdentifier :: Maybe Text

The DB cluster identifier of the DB cluster associated with the endpoint. This parameter is stored as a lowercase string.

dbClusterEndpointResourceIdentifier :: Maybe Text

A unique system-generated identifier for an endpoint. It remains the same for the whole life of the endpoint.

endpoint :: Maybe Text

The DNS address of the endpoint.

staticMembers :: Maybe [Text]

List of DB instance identifiers that are part of the custom endpoint group.

newDBClusterEndpoint :: DBClusterEndpoint

Create a value of DBClusterEndpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DBClusterEndpoint', dbClusterEndpoint_status - The current status of the endpoint. One of: creating, available, deleting, inactive, modifying. The inactive state applies to an endpoint that can't be used for a certain kind of cluster, such as a writer endpoint for a read-only secondary cluster in a global database.

$sel:dbClusterEndpointIdentifier:DBClusterEndpoint', dbClusterEndpoint_dbClusterEndpointIdentifier - The identifier associated with the endpoint. This parameter is stored as a lowercase string.

$sel:excludedMembers:DBClusterEndpoint', dbClusterEndpoint_excludedMembers - List of DB instance identifiers that aren't part of the custom endpoint group. All other eligible instances are reachable through the custom endpoint. Only relevant if the list of static members is empty.

$sel:endpointType:DBClusterEndpoint', dbClusterEndpoint_endpointType - The type of the endpoint. One of: READER, WRITER, CUSTOM.

$sel:customEndpointType:DBClusterEndpoint', dbClusterEndpoint_customEndpointType - The type associated with a custom endpoint. One of: READER, WRITER, ANY.

$sel:dbClusterEndpointArn:DBClusterEndpoint', dbClusterEndpoint_dbClusterEndpointArn - The Amazon Resource Name (ARN) for the endpoint.

$sel:dbClusterIdentifier:DBClusterEndpoint', dbClusterEndpoint_dbClusterIdentifier - The DB cluster identifier of the DB cluster associated with the endpoint. This parameter is stored as a lowercase string.

$sel:dbClusterEndpointResourceIdentifier:DBClusterEndpoint', dbClusterEndpoint_dbClusterEndpointResourceIdentifier - A unique system-generated identifier for an endpoint. It remains the same for the whole life of the endpoint.

$sel:endpoint:DBClusterEndpoint', dbClusterEndpoint_endpoint - The DNS address of the endpoint.

$sel:staticMembers:DBClusterEndpoint', dbClusterEndpoint_staticMembers - List of DB instance identifiers that are part of the custom endpoint group.

dbClusterEndpoint_status :: Lens' DBClusterEndpoint (Maybe Text)

The current status of the endpoint. One of: creating, available, deleting, inactive, modifying. The inactive state applies to an endpoint that can't be used for a certain kind of cluster, such as a writer endpoint for a read-only secondary cluster in a global database.

dbClusterEndpoint_dbClusterEndpointIdentifier :: Lens' DBClusterEndpoint (Maybe Text)

The identifier associated with the endpoint. This parameter is stored as a lowercase string.

dbClusterEndpoint_excludedMembers :: Lens' DBClusterEndpoint (Maybe [Text])

List of DB instance identifiers that aren't part of the custom endpoint group. All other eligible instances are reachable through the custom endpoint. Only relevant if the list of static members is empty.

dbClusterEndpoint_endpointType :: Lens' DBClusterEndpoint (Maybe Text)

The type of the endpoint. One of: READER, WRITER, CUSTOM.

dbClusterEndpoint_customEndpointType :: Lens' DBClusterEndpoint (Maybe Text)

The type associated with a custom endpoint. One of: READER, WRITER, ANY.

dbClusterEndpoint_dbClusterEndpointArn :: Lens' DBClusterEndpoint (Maybe Text)

The Amazon Resource Name (ARN) for the endpoint.

dbClusterEndpoint_dbClusterIdentifier :: Lens' DBClusterEndpoint (Maybe Text)

The DB cluster identifier of the DB cluster associated with the endpoint. This parameter is stored as a lowercase string.

dbClusterEndpoint_dbClusterEndpointResourceIdentifier :: Lens' DBClusterEndpoint (Maybe Text)

A unique system-generated identifier for an endpoint. It remains the same for the whole life of the endpoint.

dbClusterEndpoint_endpoint :: Lens' DBClusterEndpoint (Maybe Text)

The DNS address of the endpoint.

dbClusterEndpoint_staticMembers :: Lens' DBClusterEndpoint (Maybe [Text])

List of DB instance identifiers that are part of the custom endpoint group.