Network.AWS.RDS.Types.DBClusterRole

Description

Documentation

data DBClusterRole

Describes an Amazon Web Services Identity and Access Management (IAM) role that is associated with a DB cluster.

See: newDBClusterRole smart constructor.

Constructors

DBClusterRole'

Fields

status :: Maybe Text

Describes the state of association between the IAM role and the DB cluster. The Status property returns one of the following values:

ACTIVE - the IAM role ARN is associated with the DB cluster and can be used to access other Amazon Web Services on your behalf.

PENDING - the IAM role ARN is being associated with the DB cluster.

INVALID - the IAM role ARN is associated with the DB cluster, but the DB cluster is unable to assume the IAM role in order to access other Amazon Web Services on your behalf.

featureName :: Maybe Text

The name of the feature associated with the Amazon Web Services Identity and Access Management (IAM) role. For the list of supported feature names, see DBEngineVersion.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) of the IAM role that is associated with the DB cluster.

newDBClusterRole :: DBClusterRole

Create a value of DBClusterRole with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DBClusterRole', dbClusterRole_status - Describes the state of association between the IAM role and the DB cluster. The Status property returns one of the following values:

ACTIVE - the IAM role ARN is associated with the DB cluster and can be used to access other Amazon Web Services on your behalf.

PENDING - the IAM role ARN is being associated with the DB cluster.

INVALID - the IAM role ARN is associated with the DB cluster, but the DB cluster is unable to assume the IAM role in order to access other Amazon Web Services on your behalf.

$sel:featureName:DBClusterRole', dbClusterRole_featureName - The name of the feature associated with the Amazon Web Services Identity and Access Management (IAM) role. For the list of supported feature names, see DBEngineVersion.

$sel:roleArn:DBClusterRole', dbClusterRole_roleArn - The Amazon Resource Name (ARN) of the IAM role that is associated with the DB cluster.

dbClusterRole_status :: Lens' DBClusterRole (Maybe Text)

Describes the state of association between the IAM role and the DB cluster. The Status property returns one of the following values:

ACTIVE - the IAM role ARN is associated with the DB cluster and can be used to access other Amazon Web Services on your behalf.

PENDING - the IAM role ARN is being associated with the DB cluster.

INVALID - the IAM role ARN is associated with the DB cluster, but the DB cluster is unable to assume the IAM role in order to access other Amazon Web Services on your behalf.

dbClusterRole_featureName :: Lens' DBClusterRole (Maybe Text)

The name of the feature associated with the Amazon Web Services Identity and Access Management (IAM) role. For the list of supported feature names, see DBEngineVersion.

dbClusterRole_roleArn :: Lens' DBClusterRole (Maybe Text)

The Amazon Resource Name (ARN) of the IAM role that is associated with the DB cluster.