Network.AWS.RDS.Types.DBInstanceAutomatedBackup

Description

Documentation

data DBInstanceAutomatedBackup

An automated backup of a DB instance. It consists of system backups, transaction logs, and the database instance properties that existed at the time you deleted the source instance.

See: newDBInstanceAutomatedBackup smart constructor.

Constructors

DBInstanceAutomatedBackup'

Fields

backupRetentionPeriod :: Maybe Int

The retention period for the automated backups.

status :: Maybe Text

Provides a list of status information for an automated backup:

active - automated backups for current instances

retained - automated backups for deleted instances

creating - automated backups that are waiting for the first automated snapshot to be available.

storageType :: Maybe Text

Specifies the storage type associated with the automated backup.

dbInstanceAutomatedBackupsArn :: Maybe Text

The Amazon Resource Name (ARN) for the replicated automated backups.

encrypted :: Maybe Bool

Specifies whether the automated backup is encrypted.

optionGroupName :: Maybe Text

The option group the automated backup is associated with. If omitted, the default option group for the engine specified is used.

instanceCreateTime :: Maybe ISO8601

Provides the date and time that the DB instance was created.

dbiResourceId :: Maybe Text

The identifier for the source DB instance, which can't be changed and which is unique to an Amazon Web Services Region.

masterUsername :: Maybe Text

The license model of an automated backup.

kmsKeyId :: Maybe Text

The Amazon Web Services KMS key ID for an automated backup.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

availabilityZone :: Maybe Text

The Availability Zone that the automated backup was created in. For information on Amazon Web Services Regions and Availability Zones, see Regions and Availability Zones.

engineVersion :: Maybe Text

The version of the database engine for the automated backup.

restoreWindow :: Maybe RestoreWindow

Earliest and latest time an instance can be restored to.

licenseModel :: Maybe Text

License model information for the automated backup.

dbInstanceIdentifier :: Maybe Text

The customer id of the instance that is/was associated with the automated backup.

port :: Maybe Int

The port number that the automated backup used for connections.

Default: Inherits from the source DB instance

Valid Values: 1150-65535

engine :: Maybe Text

The name of the database engine for this automated backup.

dbInstanceAutomatedBackupsReplications :: Maybe [DBInstanceAutomatedBackupsReplication]

The list of replications to different Amazon Web Services Regions associated with the automated backup.

tdeCredentialArn :: Maybe Text

The ARN from the key store with which the automated backup is associated for TDE encryption.

timezone :: Maybe Text

The time zone of the automated backup. In most cases, the Timezone element is empty. Timezone content appears only for Microsoft SQL Server DB instances that were created with a time zone specified.

region :: Maybe Text

The Amazon Web Services Region associated with the automated backup.

vpcId :: Maybe Text

Provides the VPC ID associated with the DB instance

allocatedStorage :: Maybe Int

Specifies the allocated storage size in gibibytes (GiB).

iops :: Maybe Int

The IOPS (I/O operations per second) value for the automated backup.

iAMDatabaseAuthenticationEnabled :: Maybe Bool

True if mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts is enabled, and otherwise false.

dbInstanceArn :: Maybe Text

The Amazon Resource Name (ARN) for the automated backups.

newDBInstanceAutomatedBackup :: DBInstanceAutomatedBackup

Create a value of DBInstanceAutomatedBackup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupRetentionPeriod:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_backupRetentionPeriod - The retention period for the automated backups.

$sel:status:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_status - Provides a list of status information for an automated backup:

active - automated backups for current instances

retained - automated backups for deleted instances

creating - automated backups that are waiting for the first automated snapshot to be available.

$sel:storageType:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_storageType - Specifies the storage type associated with the automated backup.

$sel:dbInstanceAutomatedBackupsArn:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_dbInstanceAutomatedBackupsArn - The Amazon Resource Name (ARN) for the replicated automated backups.

$sel:encrypted:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_encrypted - Specifies whether the automated backup is encrypted.

$sel:optionGroupName:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_optionGroupName - The option group the automated backup is associated with. If omitted, the default option group for the engine specified is used.

$sel:instanceCreateTime:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_instanceCreateTime - Provides the date and time that the DB instance was created.

$sel:dbiResourceId:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_dbiResourceId - The identifier for the source DB instance, which can't be changed and which is unique to an Amazon Web Services Region.

$sel:masterUsername:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_masterUsername - The license model of an automated backup.

$sel:kmsKeyId:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_kmsKeyId - The Amazon Web Services KMS key ID for an automated backup.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

$sel:availabilityZone:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_availabilityZone - The Availability Zone that the automated backup was created in. For information on Amazon Web Services Regions and Availability Zones, see Regions and Availability Zones.

$sel:engineVersion:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_engineVersion - The version of the database engine for the automated backup.

$sel:restoreWindow:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_restoreWindow - Earliest and latest time an instance can be restored to.

$sel:licenseModel:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_licenseModel - License model information for the automated backup.

$sel:dbInstanceIdentifier:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_dbInstanceIdentifier - The customer id of the instance that is/was associated with the automated backup.

$sel:port:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_port - The port number that the automated backup used for connections.

Default: Inherits from the source DB instance

Valid Values: 1150-65535

$sel:engine:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_engine - The name of the database engine for this automated backup.

$sel:dbInstanceAutomatedBackupsReplications:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_dbInstanceAutomatedBackupsReplications - The list of replications to different Amazon Web Services Regions associated with the automated backup.

$sel:tdeCredentialArn:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_tdeCredentialArn - The ARN from the key store with which the automated backup is associated for TDE encryption.

$sel:timezone:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_timezone - The time zone of the automated backup. In most cases, the Timezone element is empty. Timezone content appears only for Microsoft SQL Server DB instances that were created with a time zone specified.

$sel:region:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_region - The Amazon Web Services Region associated with the automated backup.

$sel:vpcId:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_vpcId - Provides the VPC ID associated with the DB instance

$sel:allocatedStorage:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_allocatedStorage - Specifies the allocated storage size in gibibytes (GiB).

$sel:iops:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_iops - The IOPS (I/O operations per second) value for the automated backup.

$sel:iAMDatabaseAuthenticationEnabled:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_iAMDatabaseAuthenticationEnabled - True if mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts is enabled, and otherwise false.

$sel:dbInstanceArn:DBInstanceAutomatedBackup', dbInstanceAutomatedBackup_dbInstanceArn - The Amazon Resource Name (ARN) for the automated backups.

dbInstanceAutomatedBackup_backupRetentionPeriod :: Lens' DBInstanceAutomatedBackup (Maybe Int)

The retention period for the automated backups.

dbInstanceAutomatedBackup_status :: Lens' DBInstanceAutomatedBackup (Maybe Text)

Provides a list of status information for an automated backup:

active - automated backups for current instances

retained - automated backups for deleted instances

creating - automated backups that are waiting for the first automated snapshot to be available.

dbInstanceAutomatedBackup_storageType :: Lens' DBInstanceAutomatedBackup (Maybe Text)

Specifies the storage type associated with the automated backup.

dbInstanceAutomatedBackup_dbInstanceAutomatedBackupsArn :: Lens' DBInstanceAutomatedBackup (Maybe Text)

The Amazon Resource Name (ARN) for the replicated automated backups.

dbInstanceAutomatedBackup_encrypted :: Lens' DBInstanceAutomatedBackup (Maybe Bool)

Specifies whether the automated backup is encrypted.

dbInstanceAutomatedBackup_optionGroupName :: Lens' DBInstanceAutomatedBackup (Maybe Text)

The option group the automated backup is associated with. If omitted, the default option group for the engine specified is used.

dbInstanceAutomatedBackup_instanceCreateTime :: Lens' DBInstanceAutomatedBackup (Maybe UTCTime)

Provides the date and time that the DB instance was created.

dbInstanceAutomatedBackup_dbiResourceId :: Lens' DBInstanceAutomatedBackup (Maybe Text)

The identifier for the source DB instance, which can't be changed and which is unique to an Amazon Web Services Region.

dbInstanceAutomatedBackup_masterUsername :: Lens' DBInstanceAutomatedBackup (Maybe Text)

The license model of an automated backup.

dbInstanceAutomatedBackup_kmsKeyId :: Lens' DBInstanceAutomatedBackup (Maybe Text)

The Amazon Web Services KMS key ID for an automated backup.

The Amazon Web Services KMS key identifier is the key ARN, key ID, alias ARN, or alias name for the Amazon Web Services KMS customer master key (CMK).

dbInstanceAutomatedBackup_availabilityZone :: Lens' DBInstanceAutomatedBackup (Maybe Text)

The Availability Zone that the automated backup was created in. For information on Amazon Web Services Regions and Availability Zones, see Regions and Availability Zones.

dbInstanceAutomatedBackup_engineVersion :: Lens' DBInstanceAutomatedBackup (Maybe Text)

The version of the database engine for the automated backup.

dbInstanceAutomatedBackup_restoreWindow :: Lens' DBInstanceAutomatedBackup (Maybe RestoreWindow)

Earliest and latest time an instance can be restored to.

dbInstanceAutomatedBackup_licenseModel :: Lens' DBInstanceAutomatedBackup (Maybe Text)

License model information for the automated backup.

dbInstanceAutomatedBackup_dbInstanceIdentifier :: Lens' DBInstanceAutomatedBackup (Maybe Text)

The customer id of the instance that is/was associated with the automated backup.

dbInstanceAutomatedBackup_port :: Lens' DBInstanceAutomatedBackup (Maybe Int)

The port number that the automated backup used for connections.

Default: Inherits from the source DB instance

Valid Values: 1150-65535

dbInstanceAutomatedBackup_engine :: Lens' DBInstanceAutomatedBackup (Maybe Text)

The name of the database engine for this automated backup.

dbInstanceAutomatedBackup_dbInstanceAutomatedBackupsReplications :: Lens' DBInstanceAutomatedBackup (Maybe [DBInstanceAutomatedBackupsReplication])

The list of replications to different Amazon Web Services Regions associated with the automated backup.

dbInstanceAutomatedBackup_tdeCredentialArn :: Lens' DBInstanceAutomatedBackup (Maybe Text)

The ARN from the key store with which the automated backup is associated for TDE encryption.

dbInstanceAutomatedBackup_timezone :: Lens' DBInstanceAutomatedBackup (Maybe Text)

The time zone of the automated backup. In most cases, the Timezone element is empty. Timezone content appears only for Microsoft SQL Server DB instances that were created with a time zone specified.

dbInstanceAutomatedBackup_region :: Lens' DBInstanceAutomatedBackup (Maybe Text)

The Amazon Web Services Region associated with the automated backup.

dbInstanceAutomatedBackup_vpcId :: Lens' DBInstanceAutomatedBackup (Maybe Text)

Provides the VPC ID associated with the DB instance

dbInstanceAutomatedBackup_allocatedStorage :: Lens' DBInstanceAutomatedBackup (Maybe Int)

Specifies the allocated storage size in gibibytes (GiB).

dbInstanceAutomatedBackup_iops :: Lens' DBInstanceAutomatedBackup (Maybe Int)

The IOPS (I/O operations per second) value for the automated backup.

dbInstanceAutomatedBackup_iAMDatabaseAuthenticationEnabled :: Lens' DBInstanceAutomatedBackup (Maybe Bool)

True if mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts is enabled, and otherwise false.

dbInstanceAutomatedBackup_dbInstanceArn :: Lens' DBInstanceAutomatedBackup (Maybe Text)

The Amazon Resource Name (ARN) for the automated backups.