Network.AWS.RDS.Types.DBProxy

Description

Documentation

data DBProxy

The data structure representing a proxy managed by the RDS Proxy.

This data type is used as a response element in the DescribeDBProxies action.

See: newDBProxy smart constructor.

Constructors

DBProxy'

Fields

status :: Maybe DBProxyStatus

The current status of this proxy. A status of available means the proxy is ready to handle requests. Other values indicate that you must wait for the proxy to be ready, or take some action to resolve an issue.

dbProxyArn :: Maybe Text

The Amazon Resource Name (ARN) for the proxy.

debugLogging :: Maybe Bool

Whether the proxy includes detailed information about SQL statements in its logs. This information helps you to debug issues involving SQL behavior or the performance and scalability of the proxy connections. The debug information includes the text of SQL statements that you submit through the proxy. Thus, only enable this setting when needed for debugging, and only when you have security measures in place to safeguard any sensitive information that appears in the logs.

vpcSubnetIds :: Maybe [Text]

The EC2 subnet IDs for the proxy.

vpcId :: Maybe Text

Provides the VPC ID of the DB proxy.

engineFamily :: Maybe Text

The engine family applies to MySQL and PostgreSQL for both RDS and Aurora.

auth :: Maybe [UserAuthConfigInfo]

One or more data structures specifying the authorization mechanism to connect to the associated RDS DB instance or Aurora DB cluster.

requireTLS :: Maybe Bool

Indicates whether Transport Layer Security (TLS) encryption is required for connections to the proxy.

idleClientTimeout :: Maybe Int

The number of seconds a connection to the proxy can have no activity before the proxy drops the client connection. The proxy keeps the underlying database connection open and puts it back into the connection pool for reuse by later connection requests.

Default: 1800 (30 minutes)

Constraints: 1 to 28,800

updatedDate :: Maybe ISO8601

The date and time when the proxy was last updated.

createdDate :: Maybe ISO8601

The date and time when the proxy was first created.

vpcSecurityGroupIds :: Maybe [Text]

Provides a list of VPC security groups that the proxy belongs to.

dbProxyName :: Maybe Text

The identifier for the proxy. This name must be unique for all proxies owned by your Amazon Web Services account in the specified Amazon Web Services Region.

endpoint :: Maybe Text

The endpoint that you can use to connect to the DB proxy. You include the endpoint value in the connection string for a database client application.

roleArn :: Maybe Text

The Amazon Resource Name (ARN) for the IAM role that the proxy uses to access Amazon Secrets Manager.

newDBProxy :: DBProxy

Create a value of DBProxy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DBProxy', dbProxy_status - The current status of this proxy. A status of available means the proxy is ready to handle requests. Other values indicate that you must wait for the proxy to be ready, or take some action to resolve an issue.

$sel:dbProxyArn:DBProxy', dbProxy_dbProxyArn - The Amazon Resource Name (ARN) for the proxy.

$sel:debugLogging:DBProxy', dbProxy_debugLogging - Whether the proxy includes detailed information about SQL statements in its logs. This information helps you to debug issues involving SQL behavior or the performance and scalability of the proxy connections. The debug information includes the text of SQL statements that you submit through the proxy. Thus, only enable this setting when needed for debugging, and only when you have security measures in place to safeguard any sensitive information that appears in the logs.

$sel:vpcSubnetIds:DBProxy', dbProxy_vpcSubnetIds - The EC2 subnet IDs for the proxy.

$sel:vpcId:DBProxy', dbProxy_vpcId - Provides the VPC ID of the DB proxy.

$sel:engineFamily:DBProxy', dbProxy_engineFamily - The engine family applies to MySQL and PostgreSQL for both RDS and Aurora.

$sel:auth:DBProxy', dbProxy_auth - One or more data structures specifying the authorization mechanism to connect to the associated RDS DB instance or Aurora DB cluster.

$sel:requireTLS:DBProxy', dbProxy_requireTLS - Indicates whether Transport Layer Security (TLS) encryption is required for connections to the proxy.

$sel:idleClientTimeout:DBProxy', dbProxy_idleClientTimeout - The number of seconds a connection to the proxy can have no activity before the proxy drops the client connection. The proxy keeps the underlying database connection open and puts it back into the connection pool for reuse by later connection requests.

Default: 1800 (30 minutes)

Constraints: 1 to 28,800

$sel:updatedDate:DBProxy', dbProxy_updatedDate - The date and time when the proxy was last updated.

$sel:createdDate:DBProxy', dbProxy_createdDate - The date and time when the proxy was first created.

$sel:vpcSecurityGroupIds:DBProxy', dbProxy_vpcSecurityGroupIds - Provides a list of VPC security groups that the proxy belongs to.

$sel:dbProxyName:DBProxy', dbProxy_dbProxyName - The identifier for the proxy. This name must be unique for all proxies owned by your Amazon Web Services account in the specified Amazon Web Services Region.

$sel:endpoint:DBProxy', dbProxy_endpoint - The endpoint that you can use to connect to the DB proxy. You include the endpoint value in the connection string for a database client application.

$sel:roleArn:DBProxy', dbProxy_roleArn - The Amazon Resource Name (ARN) for the IAM role that the proxy uses to access Amazon Secrets Manager.

dbProxy_status :: Lens' DBProxy (Maybe DBProxyStatus)

The current status of this proxy. A status of available means the proxy is ready to handle requests. Other values indicate that you must wait for the proxy to be ready, or take some action to resolve an issue.

dbProxy_dbProxyArn :: Lens' DBProxy (Maybe Text)

The Amazon Resource Name (ARN) for the proxy.

dbProxy_debugLogging :: Lens' DBProxy (Maybe Bool)

Whether the proxy includes detailed information about SQL statements in its logs. This information helps you to debug issues involving SQL behavior or the performance and scalability of the proxy connections. The debug information includes the text of SQL statements that you submit through the proxy. Thus, only enable this setting when needed for debugging, and only when you have security measures in place to safeguard any sensitive information that appears in the logs.

dbProxy_vpcSubnetIds :: Lens' DBProxy (Maybe [Text])

The EC2 subnet IDs for the proxy.

dbProxy_vpcId :: Lens' DBProxy (Maybe Text)

Provides the VPC ID of the DB proxy.

dbProxy_engineFamily :: Lens' DBProxy (Maybe Text)

The engine family applies to MySQL and PostgreSQL for both RDS and Aurora.

dbProxy_auth :: Lens' DBProxy (Maybe [UserAuthConfigInfo])

One or more data structures specifying the authorization mechanism to connect to the associated RDS DB instance or Aurora DB cluster.

dbProxy_requireTLS :: Lens' DBProxy (Maybe Bool)

Indicates whether Transport Layer Security (TLS) encryption is required for connections to the proxy.

dbProxy_idleClientTimeout :: Lens' DBProxy (Maybe Int)

The number of seconds a connection to the proxy can have no activity before the proxy drops the client connection. The proxy keeps the underlying database connection open and puts it back into the connection pool for reuse by later connection requests.

Default: 1800 (30 minutes)

Constraints: 1 to 28,800

dbProxy_updatedDate :: Lens' DBProxy (Maybe UTCTime)

The date and time when the proxy was last updated.

dbProxy_createdDate :: Lens' DBProxy (Maybe UTCTime)

The date and time when the proxy was first created.

dbProxy_vpcSecurityGroupIds :: Lens' DBProxy (Maybe [Text])

Provides a list of VPC security groups that the proxy belongs to.

dbProxy_dbProxyName :: Lens' DBProxy (Maybe Text)

The identifier for the proxy. This name must be unique for all proxies owned by your Amazon Web Services account in the specified Amazon Web Services Region.

dbProxy_endpoint :: Lens' DBProxy (Maybe Text)

The endpoint that you can use to connect to the DB proxy. You include the endpoint value in the connection string for a database client application.

dbProxy_roleArn :: Lens' DBProxy (Maybe Text)

The Amazon Resource Name (ARN) for the IAM role that the proxy uses to access Amazon Secrets Manager.