Network.AWS.RDS.Types.DBProxyTarget

Description

Documentation

data DBProxyTarget

Contains the details for an RDS Proxy target. It represents an RDS DB instance or Aurora DB cluster that the proxy can connect to. One or more targets are associated with an RDS Proxy target group.

This data type is used as a response element in the DescribeDBProxyTargets action.

See: newDBProxyTarget smart constructor.

Constructors

DBProxyTarget'

Fields

trackedClusterId :: Maybe Text

The DB cluster identifier when the target represents an Aurora DB cluster. This field is blank when the target represents an RDS DB instance.

rdsResourceId :: Maybe Text

The identifier representing the target. It can be the instance identifier for an RDS DB instance, or the cluster identifier for an Aurora DB cluster.

targetArn :: Maybe Text

The Amazon Resource Name (ARN) for the RDS DB instance or Aurora DB cluster.

role' :: Maybe TargetRole

A value that indicates whether the target of the proxy can be used for read/write or read-only operations.

port :: Maybe Int

The port that the RDS Proxy uses to connect to the target RDS DB instance or Aurora DB cluster.

endpoint :: Maybe Text

The writer endpoint for the RDS DB instance or Aurora DB cluster.

type' :: Maybe TargetType

Specifies the kind of database, such as an RDS DB instance or an Aurora DB cluster, that the target represents.

targetHealth :: Maybe TargetHealth

Information about the connection health of the RDS Proxy target.

newDBProxyTarget :: DBProxyTarget

Create a value of DBProxyTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:trackedClusterId:DBProxyTarget', dbProxyTarget_trackedClusterId - The DB cluster identifier when the target represents an Aurora DB cluster. This field is blank when the target represents an RDS DB instance.

$sel:rdsResourceId:DBProxyTarget', dbProxyTarget_rdsResourceId - The identifier representing the target. It can be the instance identifier for an RDS DB instance, or the cluster identifier for an Aurora DB cluster.

$sel:targetArn:DBProxyTarget', dbProxyTarget_targetArn - The Amazon Resource Name (ARN) for the RDS DB instance or Aurora DB cluster.

$sel:role':DBProxyTarget', dbProxyTarget_role - A value that indicates whether the target of the proxy can be used for read/write or read-only operations.

$sel:port:DBProxyTarget', dbProxyTarget_port - The port that the RDS Proxy uses to connect to the target RDS DB instance or Aurora DB cluster.

$sel:endpoint:DBProxyTarget', dbProxyTarget_endpoint - The writer endpoint for the RDS DB instance or Aurora DB cluster.

$sel:type':DBProxyTarget', dbProxyTarget_type - Specifies the kind of database, such as an RDS DB instance or an Aurora DB cluster, that the target represents.

$sel:targetHealth:DBProxyTarget', dbProxyTarget_targetHealth - Information about the connection health of the RDS Proxy target.

dbProxyTarget_trackedClusterId :: Lens' DBProxyTarget (Maybe Text)

The DB cluster identifier when the target represents an Aurora DB cluster. This field is blank when the target represents an RDS DB instance.

dbProxyTarget_rdsResourceId :: Lens' DBProxyTarget (Maybe Text)

The identifier representing the target. It can be the instance identifier for an RDS DB instance, or the cluster identifier for an Aurora DB cluster.

dbProxyTarget_targetArn :: Lens' DBProxyTarget (Maybe Text)

The Amazon Resource Name (ARN) for the RDS DB instance or Aurora DB cluster.

dbProxyTarget_role :: Lens' DBProxyTarget (Maybe TargetRole)

A value that indicates whether the target of the proxy can be used for read/write or read-only operations.

dbProxyTarget_port :: Lens' DBProxyTarget (Maybe Int)

The port that the RDS Proxy uses to connect to the target RDS DB instance or Aurora DB cluster.

dbProxyTarget_endpoint :: Lens' DBProxyTarget (Maybe Text)

The writer endpoint for the RDS DB instance or Aurora DB cluster.

dbProxyTarget_type :: Lens' DBProxyTarget (Maybe TargetType)

Specifies the kind of database, such as an RDS DB instance or an Aurora DB cluster, that the target represents.

dbProxyTarget_targetHealth :: Lens' DBProxyTarget (Maybe TargetHealth)

Information about the connection health of the RDS Proxy target.