Network.AWS.RDS.Types.DBProxyTargetGroup

Description

Documentation

data DBProxyTargetGroup

Represents a set of RDS DB instances, Aurora DB clusters, or both that a proxy can connect to. Currently, each target group is associated with exactly one RDS DB instance or Aurora DB cluster.

This data type is used as a response element in the DescribeDBProxyTargetGroups action.

See: newDBProxyTargetGroup smart constructor.

Constructors

DBProxyTargetGroup'

Fields

status :: Maybe Text

The current status of this target group. A status of available means the target group is correctly associated with a database. Other values indicate that you must wait for the target group to be ready, or take some action to resolve an issue.

connectionPoolConfig :: Maybe ConnectionPoolConfigurationInfo

The settings that determine the size and behavior of the connection pool for the target group.

targetGroupArn :: Maybe Text

The Amazon Resource Name (ARN) representing the target group.

updatedDate :: Maybe ISO8601

The date and time when the target group was last updated.

createdDate :: Maybe ISO8601

The date and time when the target group was first created.

dbProxyName :: Maybe Text

The identifier for the RDS proxy associated with this target group.

targetGroupName :: Maybe Text

The identifier for the target group. This name must be unique for all target groups owned by your Amazon Web Services account in the specified Amazon Web Services Region.

isDefault :: Maybe Bool

Whether this target group is the first one used for connection requests by the associated proxy. Because each proxy is currently associated with a single target group, currently this setting is always true.

newDBProxyTargetGroup :: DBProxyTargetGroup

Create a value of DBProxyTargetGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DBProxyTargetGroup', dbProxyTargetGroup_status - The current status of this target group. A status of available means the target group is correctly associated with a database. Other values indicate that you must wait for the target group to be ready, or take some action to resolve an issue.

$sel:connectionPoolConfig:DBProxyTargetGroup', dbProxyTargetGroup_connectionPoolConfig - The settings that determine the size and behavior of the connection pool for the target group.

$sel:targetGroupArn:DBProxyTargetGroup', dbProxyTargetGroup_targetGroupArn - The Amazon Resource Name (ARN) representing the target group.

$sel:updatedDate:DBProxyTargetGroup', dbProxyTargetGroup_updatedDate - The date and time when the target group was last updated.

$sel:createdDate:DBProxyTargetGroup', dbProxyTargetGroup_createdDate - The date and time when the target group was first created.

$sel:dbProxyName:DBProxyTargetGroup', dbProxyTargetGroup_dbProxyName - The identifier for the RDS proxy associated with this target group.

$sel:targetGroupName:DBProxyTargetGroup', dbProxyTargetGroup_targetGroupName - The identifier for the target group. This name must be unique for all target groups owned by your Amazon Web Services account in the specified Amazon Web Services Region.

$sel:isDefault:DBProxyTargetGroup', dbProxyTargetGroup_isDefault - Whether this target group is the first one used for connection requests by the associated proxy. Because each proxy is currently associated with a single target group, currently this setting is always true.

dbProxyTargetGroup_status :: Lens' DBProxyTargetGroup (Maybe Text)

The current status of this target group. A status of available means the target group is correctly associated with a database. Other values indicate that you must wait for the target group to be ready, or take some action to resolve an issue.

dbProxyTargetGroup_connectionPoolConfig :: Lens' DBProxyTargetGroup (Maybe ConnectionPoolConfigurationInfo)

The settings that determine the size and behavior of the connection pool for the target group.

dbProxyTargetGroup_targetGroupArn :: Lens' DBProxyTargetGroup (Maybe Text)

The Amazon Resource Name (ARN) representing the target group.

dbProxyTargetGroup_updatedDate :: Lens' DBProxyTargetGroup (Maybe UTCTime)

The date and time when the target group was last updated.

dbProxyTargetGroup_createdDate :: Lens' DBProxyTargetGroup (Maybe UTCTime)

The date and time when the target group was first created.

dbProxyTargetGroup_dbProxyName :: Lens' DBProxyTargetGroup (Maybe Text)

The identifier for the RDS proxy associated with this target group.

dbProxyTargetGroup_targetGroupName :: Lens' DBProxyTargetGroup (Maybe Text)

The identifier for the target group. This name must be unique for all target groups owned by your Amazon Web Services account in the specified Amazon Web Services Region.

dbProxyTargetGroup_isDefault :: Lens' DBProxyTargetGroup (Maybe Bool)

Whether this target group is the first one used for connection requests by the associated proxy. Because each proxy is currently associated with a single target group, currently this setting is always true.