Network.AWS.RDS.Types.DBSecurityGroup

Description

Documentation

data DBSecurityGroup

Contains the details for an Amazon RDS DB security group.

This data type is used as a response element in the DescribeDBSecurityGroups action.

See: newDBSecurityGroup smart constructor.

Constructors

DBSecurityGroup'

Fields

vpcId :: Maybe Text

Provides the VpcId of the DB security group.

ownerId :: Maybe Text

Provides the Amazon Web Services ID of the owner of a specific DB security group.

dbSecurityGroupArn :: Maybe Text

The Amazon Resource Name (ARN) for the DB security group.

iPRanges :: Maybe [IPRange]

Contains a list of IPRange elements.

dbSecurityGroupName :: Maybe Text

Specifies the name of the DB security group.

eC2SecurityGroups :: Maybe [EC2SecurityGroup]

Contains a list of EC2SecurityGroup elements.

dbSecurityGroupDescription :: Maybe Text

Provides the description of the DB security group.

newDBSecurityGroup :: DBSecurityGroup

Create a value of DBSecurityGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpcId:DBSecurityGroup', dbSecurityGroup_vpcId - Provides the VpcId of the DB security group.

$sel:ownerId:DBSecurityGroup', dbSecurityGroup_ownerId - Provides the Amazon Web Services ID of the owner of a specific DB security group.

$sel:dbSecurityGroupArn:DBSecurityGroup', dbSecurityGroup_dbSecurityGroupArn - The Amazon Resource Name (ARN) for the DB security group.

$sel:iPRanges:DBSecurityGroup', dbSecurityGroup_iPRanges - Contains a list of IPRange elements.

$sel:dbSecurityGroupName:DBSecurityGroup', dbSecurityGroup_dbSecurityGroupName - Specifies the name of the DB security group.

$sel:eC2SecurityGroups:DBSecurityGroup', dbSecurityGroup_eC2SecurityGroups - Contains a list of EC2SecurityGroup elements.

$sel:dbSecurityGroupDescription:DBSecurityGroup', dbSecurityGroup_dbSecurityGroupDescription - Provides the description of the DB security group.

dbSecurityGroup_vpcId :: Lens' DBSecurityGroup (Maybe Text)

Provides the VpcId of the DB security group.

dbSecurityGroup_ownerId :: Lens' DBSecurityGroup (Maybe Text)

Provides the Amazon Web Services ID of the owner of a specific DB security group.

dbSecurityGroup_dbSecurityGroupArn :: Lens' DBSecurityGroup (Maybe Text)

The Amazon Resource Name (ARN) for the DB security group.

dbSecurityGroup_iPRanges :: Lens' DBSecurityGroup (Maybe [IPRange])

Contains a list of IPRange elements.

dbSecurityGroup_dbSecurityGroupName :: Lens' DBSecurityGroup (Maybe Text)

Specifies the name of the DB security group.

dbSecurityGroup_eC2SecurityGroups :: Lens' DBSecurityGroup (Maybe [EC2SecurityGroup])

Contains a list of EC2SecurityGroup elements.

dbSecurityGroup_dbSecurityGroupDescription :: Lens' DBSecurityGroup (Maybe Text)

Provides the description of the DB security group.