Network.AWS.RDS.Types.DBSnapshotAttribute

Description

Documentation

data DBSnapshotAttribute

Contains the name and values of a manual DB snapshot attribute

Manual DB snapshot attributes are used to authorize other Amazon Web Services accounts to restore a manual DB snapshot. For more information, see the ModifyDBSnapshotAttribute API.

See: newDBSnapshotAttribute smart constructor.

Constructors

DBSnapshotAttribute'

Fields

attributeValues :: Maybe [Text]

The value or values for the manual DB snapshot attribute.

If the AttributeName field is set to restore, then this element returns a list of IDs of the Amazon Web Services accounts that are authorized to copy or restore the manual DB snapshot. If a value of all is in the list, then the manual DB snapshot is public and available for any Amazon Web Services account to copy or restore.

attributeName :: Maybe Text

The name of the manual DB snapshot attribute.

The attribute named restore refers to the list of Amazon Web Services accounts that have permission to copy or restore the manual DB cluster snapshot. For more information, see the ModifyDBSnapshotAttribute API action.

newDBSnapshotAttribute :: DBSnapshotAttribute

Create a value of DBSnapshotAttribute with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:attributeValues:DBSnapshotAttribute', dbSnapshotAttribute_attributeValues - The value or values for the manual DB snapshot attribute.

If the AttributeName field is set to restore, then this element returns a list of IDs of the Amazon Web Services accounts that are authorized to copy or restore the manual DB snapshot. If a value of all is in the list, then the manual DB snapshot is public and available for any Amazon Web Services account to copy or restore.

$sel:attributeName:DBSnapshotAttribute', dbSnapshotAttribute_attributeName - The name of the manual DB snapshot attribute.

The attribute named restore refers to the list of Amazon Web Services accounts that have permission to copy or restore the manual DB cluster snapshot. For more information, see the ModifyDBSnapshotAttribute API action.

dbSnapshotAttribute_attributeValues :: Lens' DBSnapshotAttribute (Maybe [Text])

The value or values for the manual DB snapshot attribute.

If the AttributeName field is set to restore, then this element returns a list of IDs of the Amazon Web Services accounts that are authorized to copy or restore the manual DB snapshot. If a value of all is in the list, then the manual DB snapshot is public and available for any Amazon Web Services account to copy or restore.

dbSnapshotAttribute_attributeName :: Lens' DBSnapshotAttribute (Maybe Text)

The name of the manual DB snapshot attribute.

The attribute named restore refers to the list of Amazon Web Services accounts that have permission to copy or restore the manual DB cluster snapshot. For more information, see the ModifyDBSnapshotAttribute API action.