Network.AWS.RDS.Types.Endpoint

Description

Documentation

data Endpoint

This data type represents the information you need to connect to an Amazon RDS DB instance. This data type is used as a response element in the following actions:

CreateDBInstance

DescribeDBInstances

DeleteDBInstance

For the data structure that represents Amazon Aurora DB cluster endpoints, see DBClusterEndpoint.

See: newEndpoint smart constructor.

Constructors

Endpoint'

Fields

address :: Maybe Text

Specifies the DNS address of the DB instance.

hostedZoneId :: Maybe Text

Specifies the ID that Amazon Route 53 assigns when you create a hosted zone.

port :: Maybe Int

Specifies the port that the database engine is listening on.

newEndpoint :: Endpoint

Create a value of Endpoint with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:address:Endpoint', endpoint_address - Specifies the DNS address of the DB instance.

$sel:hostedZoneId:Endpoint', endpoint_hostedZoneId - Specifies the ID that Amazon Route 53 assigns when you create a hosted zone.

$sel:port:Endpoint', endpoint_port - Specifies the port that the database engine is listening on.

endpoint_address :: Lens' Endpoint (Maybe Text)

Specifies the DNS address of the DB instance.

endpoint_hostedZoneId :: Lens' Endpoint (Maybe Text)

Specifies the ID that Amazon Route 53 assigns when you create a hosted zone.

endpoint_port :: Lens' Endpoint (Maybe Int)

Specifies the port that the database engine is listening on.