Network.AWS.RDS.Types.EngineDefaults

Description

Documentation

data EngineDefaults

Contains the result of a successful invocation of the DescribeEngineDefaultParameters action.

See: newEngineDefaults smart constructor.

Constructors

EngineDefaults'

Fields

dbParameterGroupFamily :: Maybe Text

Specifies the name of the DB parameter group family that the engine default parameters apply to.

marker :: Maybe Text

An optional pagination token provided by a previous EngineDefaults request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

parameters :: Maybe [Parameter]

Contains a list of engine default parameters.

newEngineDefaults :: EngineDefaults

Create a value of EngineDefaults with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbParameterGroupFamily:EngineDefaults', engineDefaults_dbParameterGroupFamily - Specifies the name of the DB parameter group family that the engine default parameters apply to.

$sel:marker:EngineDefaults', engineDefaults_marker - An optional pagination token provided by a previous EngineDefaults request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

$sel:parameters:EngineDefaults', engineDefaults_parameters - Contains a list of engine default parameters.

engineDefaults_dbParameterGroupFamily :: Lens' EngineDefaults (Maybe Text)

Specifies the name of the DB parameter group family that the engine default parameters apply to.

engineDefaults_marker :: Lens' EngineDefaults (Maybe Text)

An optional pagination token provided by a previous EngineDefaults request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

engineDefaults_parameters :: Lens' EngineDefaults (Maybe [Parameter])

Contains a list of engine default parameters.