Network.AWS.RDS.Types.ExportTask

Description

Documentation

data ExportTask

Contains the details of a snapshot export to Amazon S3.

This data type is used as a response element in the DescribeExportTasks action.

See: newExportTask smart constructor.

Constructors

ExportTask'

Fields

taskEndTime :: Maybe ISO8601

The time that the snapshot export task completed.

status :: Maybe Text

The progress status of the export task.

iamRoleArn :: Maybe Text

The name of the IAM role that is used to write to Amazon S3 when exporting a snapshot.

totalExtractedDataInGB :: Maybe Int

The total amount of data exported, in gigabytes.

warningMessage :: Maybe Text

A warning about the snapshot export task.

snapshotTime :: Maybe ISO8601

The time that the snapshot was created.

s3Bucket :: Maybe Text

The Amazon S3 bucket that the snapshot is exported to.

exportOnly :: Maybe [Text]

The data exported from the snapshot. Valid values are the following:

database - Export all the data from a specified database.

database.table table-name - Export a table of the snapshot. This format is valid only for RDS for MySQL, RDS for MariaDB, and Aurora MySQL.

database.schema schema-name - Export a database schema of the snapshot. This format is valid only for RDS for PostgreSQL and Aurora PostgreSQL.

database.schema.table table-name - Export a table of the database schema. This format is valid only for RDS for PostgreSQL and Aurora PostgreSQL.

kmsKeyId :: Maybe Text

The key identifier of the Amazon Web Services KMS customer master key (CMK) that is used to encrypt the snapshot when it's exported to Amazon S3. The Amazon Web Services KMS CMK identifier is its key ARN, key ID, alias ARN, or alias name. The IAM role used for the snapshot export must have encryption and decryption permissions to use this Amazon Web Services KMS CMK.

failureCause :: Maybe Text

The reason the export failed, if it failed.

percentProgress :: Maybe Int

The progress of the snapshot export task as a percentage.

sourceArn :: Maybe Text

The Amazon Resource Name (ARN) of the snapshot exported to Amazon S3.

s3Prefix :: Maybe Text

The Amazon S3 bucket prefix that is the file name and path of the exported snapshot.

taskStartTime :: Maybe ISO8601

The time that the snapshot export task started.

exportTaskIdentifier :: Maybe Text

A unique identifier for the snapshot export task. This ID isn't an identifier for the Amazon S3 bucket where the snapshot is exported to.

newExportTask :: ExportTask

Create a value of ExportTask with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taskEndTime:ExportTask', exportTask_taskEndTime - The time that the snapshot export task completed.

$sel:status:ExportTask', exportTask_status - The progress status of the export task.

$sel:iamRoleArn:ExportTask', exportTask_iamRoleArn - The name of the IAM role that is used to write to Amazon S3 when exporting a snapshot.

$sel:totalExtractedDataInGB:ExportTask', exportTask_totalExtractedDataInGB - The total amount of data exported, in gigabytes.

$sel:warningMessage:ExportTask', exportTask_warningMessage - A warning about the snapshot export task.

$sel:snapshotTime:ExportTask', exportTask_snapshotTime - The time that the snapshot was created.

$sel:s3Bucket:ExportTask', exportTask_s3Bucket - The Amazon S3 bucket that the snapshot is exported to.

$sel:exportOnly:ExportTask', exportTask_exportOnly - The data exported from the snapshot. Valid values are the following:

database - Export all the data from a specified database.

database.table table-name - Export a table of the snapshot. This format is valid only for RDS for MySQL, RDS for MariaDB, and Aurora MySQL.

database.schema schema-name - Export a database schema of the snapshot. This format is valid only for RDS for PostgreSQL and Aurora PostgreSQL.

database.schema.table table-name - Export a table of the database schema. This format is valid only for RDS for PostgreSQL and Aurora PostgreSQL.

$sel:kmsKeyId:ExportTask', exportTask_kmsKeyId - The key identifier of the Amazon Web Services KMS customer master key (CMK) that is used to encrypt the snapshot when it's exported to Amazon S3. The Amazon Web Services KMS CMK identifier is its key ARN, key ID, alias ARN, or alias name. The IAM role used for the snapshot export must have encryption and decryption permissions to use this Amazon Web Services KMS CMK.

$sel:failureCause:ExportTask', exportTask_failureCause - The reason the export failed, if it failed.

$sel:percentProgress:ExportTask', exportTask_percentProgress - The progress of the snapshot export task as a percentage.

$sel:sourceArn:ExportTask', exportTask_sourceArn - The Amazon Resource Name (ARN) of the snapshot exported to Amazon S3.

$sel:s3Prefix:ExportTask', exportTask_s3Prefix - The Amazon S3 bucket prefix that is the file name and path of the exported snapshot.

$sel:taskStartTime:ExportTask', exportTask_taskStartTime - The time that the snapshot export task started.

$sel:exportTaskIdentifier:ExportTask', exportTask_exportTaskIdentifier - A unique identifier for the snapshot export task. This ID isn't an identifier for the Amazon S3 bucket where the snapshot is exported to.

exportTask_taskEndTime :: Lens' ExportTask (Maybe UTCTime)

The time that the snapshot export task completed.

exportTask_status :: Lens' ExportTask (Maybe Text)

The progress status of the export task.

exportTask_iamRoleArn :: Lens' ExportTask (Maybe Text)

The name of the IAM role that is used to write to Amazon S3 when exporting a snapshot.

exportTask_totalExtractedDataInGB :: Lens' ExportTask (Maybe Int)

The total amount of data exported, in gigabytes.

exportTask_warningMessage :: Lens' ExportTask (Maybe Text)

A warning about the snapshot export task.

exportTask_snapshotTime :: Lens' ExportTask (Maybe UTCTime)

The time that the snapshot was created.

exportTask_s3Bucket :: Lens' ExportTask (Maybe Text)

The Amazon S3 bucket that the snapshot is exported to.

exportTask_exportOnly :: Lens' ExportTask (Maybe [Text])

The data exported from the snapshot. Valid values are the following:

database - Export all the data from a specified database.

database.table table-name - Export a table of the snapshot. This format is valid only for RDS for MySQL, RDS for MariaDB, and Aurora MySQL.

database.schema schema-name - Export a database schema of the snapshot. This format is valid only for RDS for PostgreSQL and Aurora PostgreSQL.

database.schema.table table-name - Export a table of the database schema. This format is valid only for RDS for PostgreSQL and Aurora PostgreSQL.

exportTask_kmsKeyId :: Lens' ExportTask (Maybe Text)

The key identifier of the Amazon Web Services KMS customer master key (CMK) that is used to encrypt the snapshot when it's exported to Amazon S3. The Amazon Web Services KMS CMK identifier is its key ARN, key ID, alias ARN, or alias name. The IAM role used for the snapshot export must have encryption and decryption permissions to use this Amazon Web Services KMS CMK.

exportTask_failureCause :: Lens' ExportTask (Maybe Text)

The reason the export failed, if it failed.

exportTask_percentProgress :: Lens' ExportTask (Maybe Int)

The progress of the snapshot export task as a percentage.

exportTask_sourceArn :: Lens' ExportTask (Maybe Text)

The Amazon Resource Name (ARN) of the snapshot exported to Amazon S3.

exportTask_s3Prefix :: Lens' ExportTask (Maybe Text)

The Amazon S3 bucket prefix that is the file name and path of the exported snapshot.

exportTask_taskStartTime :: Lens' ExportTask (Maybe UTCTime)

The time that the snapshot export task started.

exportTask_exportTaskIdentifier :: Lens' ExportTask (Maybe Text)

A unique identifier for the snapshot export task. This ID isn't an identifier for the Amazon S3 bucket where the snapshot is exported to.