Network.AWS.RDS.Types.GlobalCluster

Description

Documentation

data GlobalCluster

A data type representing an Aurora global database.

See: newGlobalCluster smart constructor.

Constructors

GlobalCluster'

Fields

engineVersion :: Maybe Text

Indicates the database engine version.

status :: Maybe Text

Specifies the current state of this global database cluster.

deletionProtection :: Maybe Bool

The deletion protection setting for the new global database cluster.

storageEncrypted :: Maybe Bool

The storage encryption setting for the global database cluster.

globalClusterIdentifier :: Maybe Text

Contains a user-supplied global database cluster identifier. This identifier is the unique key that identifies a global database cluster.

engine :: Maybe Text

The Aurora database engine used by the global database cluster.

globalClusterArn :: Maybe Text

The Amazon Resource Name (ARN) for the global database cluster.

databaseName :: Maybe Text

The default database name within the new global database cluster.

globalClusterMembers :: Maybe [GlobalClusterMember]

The list of cluster IDs for secondary clusters within the global database cluster. Currently limited to 1 item.

failoverState :: Maybe FailoverState

A data object containing all properties for the current state of an in-process or pending failover process for this Aurora global database. This object is empty unless the FailoverGlobalCluster API operation has been called on this Aurora global database (GlobalCluster).

globalClusterResourceId :: Maybe Text

The Amazon Web Services Region-unique, immutable identifier for the global database cluster. This identifier is found in Amazon Web Services CloudTrail log entries whenever the Amazon Web Services KMS customer master key (CMK) for the DB cluster is accessed.

newGlobalCluster :: GlobalCluster

Create a value of GlobalCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:GlobalCluster', globalCluster_engineVersion - Indicates the database engine version.

$sel:status:GlobalCluster', globalCluster_status - Specifies the current state of this global database cluster.

$sel:deletionProtection:GlobalCluster', globalCluster_deletionProtection - The deletion protection setting for the new global database cluster.

$sel:storageEncrypted:GlobalCluster', globalCluster_storageEncrypted - The storage encryption setting for the global database cluster.

$sel:globalClusterIdentifier:GlobalCluster', globalCluster_globalClusterIdentifier - Contains a user-supplied global database cluster identifier. This identifier is the unique key that identifies a global database cluster.

$sel:engine:GlobalCluster', globalCluster_engine - The Aurora database engine used by the global database cluster.

$sel:globalClusterArn:GlobalCluster', globalCluster_globalClusterArn - The Amazon Resource Name (ARN) for the global database cluster.

$sel:databaseName:GlobalCluster', globalCluster_databaseName - The default database name within the new global database cluster.

$sel:globalClusterMembers:GlobalCluster', globalCluster_globalClusterMembers - The list of cluster IDs for secondary clusters within the global database cluster. Currently limited to 1 item.

$sel:failoverState:GlobalCluster', globalCluster_failoverState - A data object containing all properties for the current state of an in-process or pending failover process for this Aurora global database. This object is empty unless the FailoverGlobalCluster API operation has been called on this Aurora global database (GlobalCluster).

$sel:globalClusterResourceId:GlobalCluster', globalCluster_globalClusterResourceId - The Amazon Web Services Region-unique, immutable identifier for the global database cluster. This identifier is found in Amazon Web Services CloudTrail log entries whenever the Amazon Web Services KMS customer master key (CMK) for the DB cluster is accessed.

globalCluster_engineVersion :: Lens' GlobalCluster (Maybe Text)

Indicates the database engine version.

globalCluster_status :: Lens' GlobalCluster (Maybe Text)

Specifies the current state of this global database cluster.

globalCluster_deletionProtection :: Lens' GlobalCluster (Maybe Bool)

The deletion protection setting for the new global database cluster.

globalCluster_storageEncrypted :: Lens' GlobalCluster (Maybe Bool)

The storage encryption setting for the global database cluster.

globalCluster_globalClusterIdentifier :: Lens' GlobalCluster (Maybe Text)

Contains a user-supplied global database cluster identifier. This identifier is the unique key that identifies a global database cluster.

globalCluster_engine :: Lens' GlobalCluster (Maybe Text)

The Aurora database engine used by the global database cluster.

globalCluster_globalClusterArn :: Lens' GlobalCluster (Maybe Text)

The Amazon Resource Name (ARN) for the global database cluster.

globalCluster_databaseName :: Lens' GlobalCluster (Maybe Text)

The default database name within the new global database cluster.

globalCluster_globalClusterMembers :: Lens' GlobalCluster (Maybe [GlobalClusterMember])

The list of cluster IDs for secondary clusters within the global database cluster. Currently limited to 1 item.

globalCluster_failoverState :: Lens' GlobalCluster (Maybe FailoverState)

A data object containing all properties for the current state of an in-process or pending failover process for this Aurora global database. This object is empty unless the FailoverGlobalCluster API operation has been called on this Aurora global database (GlobalCluster).

globalCluster_globalClusterResourceId :: Lens' GlobalCluster (Maybe Text)

The Amazon Web Services Region-unique, immutable identifier for the global database cluster. This identifier is found in Amazon Web Services CloudTrail log entries whenever the Amazon Web Services KMS customer master key (CMK) for the DB cluster is accessed.