Network.AWS.RDS.Types.InstallationMedia

Description

Documentation

data InstallationMedia

Contains the installation media for a DB engine that requires an on-premises customer provided license, such as Microsoft SQL Server.

See: newInstallationMedia smart constructor.

Constructors

InstallationMedia'

Fields

engineVersion :: Maybe Text

The engine version of the DB engine.

status :: Maybe Text

The status of the installation medium.

installationMediaId :: Maybe Text

The installation medium ID.

engineInstallationMediaPath :: Maybe Text

The path to the installation medium for the DB engine.

engine :: Maybe Text

The DB engine.

oSInstallationMediaPath :: Maybe Text

The path to the installation medium for the operating system associated with the DB engine.

customAvailabilityZoneId :: Maybe Text

The custom Availability Zone (AZ) that contains the installation media.

failureCause :: Maybe InstallationMediaFailureCause

If an installation media failure occurred, the cause of the failure.

newInstallationMedia :: InstallationMedia

Create a value of InstallationMedia with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:InstallationMedia', installationMedia_engineVersion - The engine version of the DB engine.

$sel:status:InstallationMedia', installationMedia_status - The status of the installation medium.

$sel:installationMediaId:InstallationMedia', installationMedia_installationMediaId - The installation medium ID.

$sel:engineInstallationMediaPath:InstallationMedia', installationMedia_engineInstallationMediaPath - The path to the installation medium for the DB engine.

$sel:engine:InstallationMedia', installationMedia_engine - The DB engine.

$sel:oSInstallationMediaPath:InstallationMedia', installationMedia_oSInstallationMediaPath - The path to the installation medium for the operating system associated with the DB engine.

$sel:customAvailabilityZoneId:InstallationMedia', installationMedia_customAvailabilityZoneId - The custom Availability Zone (AZ) that contains the installation media.

$sel:failureCause:InstallationMedia', installationMedia_failureCause - If an installation media failure occurred, the cause of the failure.

installationMedia_engineVersion :: Lens' InstallationMedia (Maybe Text)

The engine version of the DB engine.

installationMedia_status :: Lens' InstallationMedia (Maybe Text)

The status of the installation medium.

installationMedia_installationMediaId :: Lens' InstallationMedia (Maybe Text)

The installation medium ID.

installationMedia_engineInstallationMediaPath :: Lens' InstallationMedia (Maybe Text)

The path to the installation medium for the DB engine.

installationMedia_engine :: Lens' InstallationMedia (Maybe Text)

The DB engine.

installationMedia_oSInstallationMediaPath :: Lens' InstallationMedia (Maybe Text)

The path to the installation medium for the operating system associated with the DB engine.

installationMedia_customAvailabilityZoneId :: Lens' InstallationMedia (Maybe Text)

The custom Availability Zone (AZ) that contains the installation media.

installationMedia_failureCause :: Lens' InstallationMedia (Maybe InstallationMediaFailureCause)

If an installation media failure occurred, the cause of the failure.