Network.AWS.RDS.Types.Option

Description

Documentation

data Option

Option details.

See: newOption smart constructor.

Constructors

Option'

Fields

optionName :: Maybe Text

The name of the option.

optionVersion :: Maybe Text

The version of the option.

dbSecurityGroupMemberships :: Maybe [DBSecurityGroupMembership]

If the option requires access to a port, then this DB security group allows access to the port.

optionDescription :: Maybe Text

The description of the option.

port :: Maybe Int

If required, the port configured for this option to use.

optionSettings :: Maybe [OptionSetting]

The option settings for this option.

persistent :: Maybe Bool

Indicate if this option is persistent.

vpcSecurityGroupMemberships :: Maybe [VpcSecurityGroupMembership]

If the option requires access to a port, then this VPC security group allows access to the port.

permanent :: Maybe Bool

Indicate if this option is permanent.

newOption :: Option

Create a value of Option with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:optionName:Option', option_optionName - The name of the option.

$sel:optionVersion:Option', option_optionVersion - The version of the option.

$sel:dbSecurityGroupMemberships:Option', option_dbSecurityGroupMemberships - If the option requires access to a port, then this DB security group allows access to the port.

$sel:optionDescription:Option', option_optionDescription - The description of the option.

$sel:port:Option', option_port - If required, the port configured for this option to use.

$sel:optionSettings:Option', option_optionSettings - The option settings for this option.

$sel:persistent:Option', option_persistent - Indicate if this option is persistent.

$sel:vpcSecurityGroupMemberships:Option', option_vpcSecurityGroupMemberships - If the option requires access to a port, then this VPC security group allows access to the port.

$sel:permanent:Option', option_permanent - Indicate if this option is permanent.

option_optionName :: Lens' Option (Maybe Text)

The name of the option.

option_optionVersion :: Lens' Option (Maybe Text)

The version of the option.

option_dbSecurityGroupMemberships :: Lens' Option (Maybe [DBSecurityGroupMembership])

If the option requires access to a port, then this DB security group allows access to the port.

option_optionDescription :: Lens' Option (Maybe Text)

The description of the option.

option_port :: Lens' Option (Maybe Int)

If required, the port configured for this option to use.

option_optionSettings :: Lens' Option (Maybe [OptionSetting])

The option settings for this option.

option_persistent :: Lens' Option (Maybe Bool)

Indicate if this option is persistent.

option_vpcSecurityGroupMemberships :: Lens' Option (Maybe [VpcSecurityGroupMembership])

If the option requires access to a port, then this VPC security group allows access to the port.

option_permanent :: Lens' Option (Maybe Bool)

Indicate if this option is permanent.