Network.AWS.RDS.Types.OptionGroup

Description

Documentation

data OptionGroup

See: newOptionGroup smart constructor.

Constructors

OptionGroup'

Fields

optionGroupArn :: Maybe Text

The Amazon Resource Name (ARN) for the option group.

engineName :: Maybe Text

Indicates the name of the engine that this option group can be applied to.

allowsVpcAndNonVpcInstanceMemberships :: Maybe Bool

Indicates whether this option group can be applied to both VPC and non-VPC instances. The value true indicates the option group can be applied to both VPC and non-VPC instances.

optionGroupName :: Maybe Text

Specifies the name of the option group.

options :: Maybe [Option]

Indicates what options are available in the option group.

optionGroupDescription :: Maybe Text

Provides a description of the option group.

majorEngineVersion :: Maybe Text

Indicates the major engine version associated with this option group.

vpcId :: Maybe Text

If AllowsVpcAndNonVpcInstanceMemberships is false, this field is blank. If AllowsVpcAndNonVpcInstanceMemberships is true and this field is blank, then this option group can be applied to both VPC and non-VPC instances. If this field contains a value, then this option group can only be applied to instances that are in the VPC indicated by this field.

newOptionGroup :: OptionGroup

Create a value of OptionGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:optionGroupArn:OptionGroup', optionGroup_optionGroupArn - The Amazon Resource Name (ARN) for the option group.

$sel:engineName:OptionGroup', optionGroup_engineName - Indicates the name of the engine that this option group can be applied to.

$sel:allowsVpcAndNonVpcInstanceMemberships:OptionGroup', optionGroup_allowsVpcAndNonVpcInstanceMemberships - Indicates whether this option group can be applied to both VPC and non-VPC instances. The value true indicates the option group can be applied to both VPC and non-VPC instances.

$sel:optionGroupName:OptionGroup', optionGroup_optionGroupName - Specifies the name of the option group.

$sel:options:OptionGroup', optionGroup_options - Indicates what options are available in the option group.

$sel:optionGroupDescription:OptionGroup', optionGroup_optionGroupDescription - Provides a description of the option group.

$sel:majorEngineVersion:OptionGroup', optionGroup_majorEngineVersion - Indicates the major engine version associated with this option group.

$sel:vpcId:OptionGroup', optionGroup_vpcId - If AllowsVpcAndNonVpcInstanceMemberships is false, this field is blank. If AllowsVpcAndNonVpcInstanceMemberships is true and this field is blank, then this option group can be applied to both VPC and non-VPC instances. If this field contains a value, then this option group can only be applied to instances that are in the VPC indicated by this field.

optionGroup_optionGroupArn :: Lens' OptionGroup (Maybe Text)

The Amazon Resource Name (ARN) for the option group.

optionGroup_engineName :: Lens' OptionGroup (Maybe Text)

Indicates the name of the engine that this option group can be applied to.

optionGroup_allowsVpcAndNonVpcInstanceMemberships :: Lens' OptionGroup (Maybe Bool)

Indicates whether this option group can be applied to both VPC and non-VPC instances. The value true indicates the option group can be applied to both VPC and non-VPC instances.

optionGroup_optionGroupName :: Lens' OptionGroup (Maybe Text)

Specifies the name of the option group.

optionGroup_options :: Lens' OptionGroup (Maybe [Option])

Indicates what options are available in the option group.

optionGroup_optionGroupDescription :: Lens' OptionGroup (Maybe Text)

Provides a description of the option group.

optionGroup_majorEngineVersion :: Lens' OptionGroup (Maybe Text)

Indicates the major engine version associated with this option group.

optionGroup_vpcId :: Lens' OptionGroup (Maybe Text)

If AllowsVpcAndNonVpcInstanceMemberships is false, this field is blank. If AllowsVpcAndNonVpcInstanceMemberships is true and this field is blank, then this option group can be applied to both VPC and non-VPC instances. If this field contains a value, then this option group can only be applied to instances that are in the VPC indicated by this field.