Network.AWS.RDS.Types.OptionGroupOption

Description

Documentation

data OptionGroupOption

Available option.

See: newOptionGroupOption smart constructor.

Constructors

OptionGroupOption'

Fields

minimumRequiredMinorEngineVersion :: Maybe Text

The minimum required engine version for the option to be applied.

optionsConflictsWith :: Maybe [Text]

The options that conflict with this option.

permanent :: Maybe Bool

Permanent options can never be removed from an option group. An option group containing a permanent option can't be removed from a DB instance.

persistent :: Maybe Bool

Persistent options can't be removed from an option group while DB instances are associated with the option group. If you disassociate all DB instances from the option group, your can remove the persistent option from the option group.

optionGroupOptionVersions :: Maybe [OptionVersion]

The versions that are available for the option.

engineName :: Maybe Text

The name of the engine that this option can be applied to.

majorEngineVersion :: Maybe Text

Indicates the major engine version that the option is available for.

name :: Maybe Text

The name of the option.

supportsOptionVersionDowngrade :: Maybe Bool

If true, you can change the option to an earlier version of the option. This only applies to options that have different versions available.

defaultPort :: Maybe Int

If the option requires a port, specifies the default port for the option.

optionGroupOptionSettings :: Maybe [OptionGroupOptionSetting]

The option settings that are available (and the default value) for each option in an option group.

requiresAutoMinorEngineVersionUpgrade :: Maybe Bool

If true, you must enable the Auto Minor Version Upgrade setting for your DB instance before you can use this option. You can enable Auto Minor Version Upgrade when you first create your DB instance, or by modifying your DB instance later.

portRequired :: Maybe Bool

Specifies whether the option requires a port.

description :: Maybe Text

The description of the option.

optionsDependedOn :: Maybe [Text]

The options that are prerequisites for this option.

vpcOnly :: Maybe Bool

If true, you can only use this option with a DB instance that is in a VPC.

newOptionGroupOption :: OptionGroupOption

Create a value of OptionGroupOption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:minimumRequiredMinorEngineVersion:OptionGroupOption', optionGroupOption_minimumRequiredMinorEngineVersion - The minimum required engine version for the option to be applied.

$sel:optionsConflictsWith:OptionGroupOption', optionGroupOption_optionsConflictsWith - The options that conflict with this option.

$sel:permanent:OptionGroupOption', optionGroupOption_permanent - Permanent options can never be removed from an option group. An option group containing a permanent option can't be removed from a DB instance.

$sel:persistent:OptionGroupOption', optionGroupOption_persistent - Persistent options can't be removed from an option group while DB instances are associated with the option group. If you disassociate all DB instances from the option group, your can remove the persistent option from the option group.

$sel:optionGroupOptionVersions:OptionGroupOption', optionGroupOption_optionGroupOptionVersions - The versions that are available for the option.

$sel:engineName:OptionGroupOption', optionGroupOption_engineName - The name of the engine that this option can be applied to.

$sel:majorEngineVersion:OptionGroupOption', optionGroupOption_majorEngineVersion - Indicates the major engine version that the option is available for.

$sel:name:OptionGroupOption', optionGroupOption_name - The name of the option.

$sel:supportsOptionVersionDowngrade:OptionGroupOption', optionGroupOption_supportsOptionVersionDowngrade - If true, you can change the option to an earlier version of the option. This only applies to options that have different versions available.

$sel:defaultPort:OptionGroupOption', optionGroupOption_defaultPort - If the option requires a port, specifies the default port for the option.

$sel:optionGroupOptionSettings:OptionGroupOption', optionGroupOption_optionGroupOptionSettings - The option settings that are available (and the default value) for each option in an option group.

$sel:requiresAutoMinorEngineVersionUpgrade:OptionGroupOption', optionGroupOption_requiresAutoMinorEngineVersionUpgrade - If true, you must enable the Auto Minor Version Upgrade setting for your DB instance before you can use this option. You can enable Auto Minor Version Upgrade when you first create your DB instance, or by modifying your DB instance later.

$sel:portRequired:OptionGroupOption', optionGroupOption_portRequired - Specifies whether the option requires a port.

$sel:description:OptionGroupOption', optionGroupOption_description - The description of the option.

$sel:optionsDependedOn:OptionGroupOption', optionGroupOption_optionsDependedOn - The options that are prerequisites for this option.

$sel:vpcOnly:OptionGroupOption', optionGroupOption_vpcOnly - If true, you can only use this option with a DB instance that is in a VPC.

optionGroupOption_minimumRequiredMinorEngineVersion :: Lens' OptionGroupOption (Maybe Text)

The minimum required engine version for the option to be applied.

optionGroupOption_optionsConflictsWith :: Lens' OptionGroupOption (Maybe [Text])

The options that conflict with this option.

optionGroupOption_permanent :: Lens' OptionGroupOption (Maybe Bool)

Permanent options can never be removed from an option group. An option group containing a permanent option can't be removed from a DB instance.

optionGroupOption_persistent :: Lens' OptionGroupOption (Maybe Bool)

Persistent options can't be removed from an option group while DB instances are associated with the option group. If you disassociate all DB instances from the option group, your can remove the persistent option from the option group.

optionGroupOption_optionGroupOptionVersions :: Lens' OptionGroupOption (Maybe [OptionVersion])

The versions that are available for the option.

optionGroupOption_engineName :: Lens' OptionGroupOption (Maybe Text)

The name of the engine that this option can be applied to.

optionGroupOption_majorEngineVersion :: Lens' OptionGroupOption (Maybe Text)

Indicates the major engine version that the option is available for.

optionGroupOption_name :: Lens' OptionGroupOption (Maybe Text)

The name of the option.

optionGroupOption_supportsOptionVersionDowngrade :: Lens' OptionGroupOption (Maybe Bool)

If true, you can change the option to an earlier version of the option. This only applies to options that have different versions available.

optionGroupOption_defaultPort :: Lens' OptionGroupOption (Maybe Int)

If the option requires a port, specifies the default port for the option.

optionGroupOption_optionGroupOptionSettings :: Lens' OptionGroupOption (Maybe [OptionGroupOptionSetting])

The option settings that are available (and the default value) for each option in an option group.

optionGroupOption_requiresAutoMinorEngineVersionUpgrade :: Lens' OptionGroupOption (Maybe Bool)

If true, you must enable the Auto Minor Version Upgrade setting for your DB instance before you can use this option. You can enable Auto Minor Version Upgrade when you first create your DB instance, or by modifying your DB instance later.

optionGroupOption_portRequired :: Lens' OptionGroupOption (Maybe Bool)

Specifies whether the option requires a port.

optionGroupOption_description :: Lens' OptionGroupOption (Maybe Text)

The description of the option.

optionGroupOption_optionsDependedOn :: Lens' OptionGroupOption (Maybe [Text])

The options that are prerequisites for this option.

optionGroupOption_vpcOnly :: Lens' OptionGroupOption (Maybe Bool)

If true, you can only use this option with a DB instance that is in a VPC.