Network.AWS.RDS.Types.OptionGroupOptionSetting

Description

Documentation

data OptionGroupOptionSetting

Option group option settings are used to display settings available for each option with their default values and other information. These values are used with the DescribeOptionGroupOptions action.

See: newOptionGroupOptionSetting smart constructor.

Constructors

OptionGroupOptionSetting'

Fields

applyType :: Maybe Text

The DB engine specific parameter type for the option group option.

minimumEngineVersionPerAllowedValue :: Maybe [MinimumEngineVersionPerAllowedValue]

The minimum DB engine version required for the corresponding allowed value for this option setting.

settingName :: Maybe Text

The name of the option group option.

defaultValue :: Maybe Text

The default value for the option group option.

isModifiable :: Maybe Bool

Boolean value where true indicates that this option group option can be changed from the default value.

settingDescription :: Maybe Text

The description of the option group option.

allowedValues :: Maybe Text

Indicates the acceptable values for the option group option.

isRequired :: Maybe Bool

Boolean value where true indicates that a value must be specified for this option setting of the option group option.

newOptionGroupOptionSetting :: OptionGroupOptionSetting

Create a value of OptionGroupOptionSetting with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:applyType:OptionGroupOptionSetting', optionGroupOptionSetting_applyType - The DB engine specific parameter type for the option group option.

$sel:minimumEngineVersionPerAllowedValue:OptionGroupOptionSetting', optionGroupOptionSetting_minimumEngineVersionPerAllowedValue - The minimum DB engine version required for the corresponding allowed value for this option setting.

$sel:settingName:OptionGroupOptionSetting', optionGroupOptionSetting_settingName - The name of the option group option.

$sel:defaultValue:OptionGroupOptionSetting', optionGroupOptionSetting_defaultValue - The default value for the option group option.

$sel:isModifiable:OptionGroupOptionSetting', optionGroupOptionSetting_isModifiable - Boolean value where true indicates that this option group option can be changed from the default value.

$sel:settingDescription:OptionGroupOptionSetting', optionGroupOptionSetting_settingDescription - The description of the option group option.

$sel:allowedValues:OptionGroupOptionSetting', optionGroupOptionSetting_allowedValues - Indicates the acceptable values for the option group option.

$sel:isRequired:OptionGroupOptionSetting', optionGroupOptionSetting_isRequired - Boolean value where true indicates that a value must be specified for this option setting of the option group option.

optionGroupOptionSetting_applyType :: Lens' OptionGroupOptionSetting (Maybe Text)

The DB engine specific parameter type for the option group option.

optionGroupOptionSetting_minimumEngineVersionPerAllowedValue :: Lens' OptionGroupOptionSetting (Maybe [MinimumEngineVersionPerAllowedValue])

The minimum DB engine version required for the corresponding allowed value for this option setting.

optionGroupOptionSetting_settingName :: Lens' OptionGroupOptionSetting (Maybe Text)

The name of the option group option.

optionGroupOptionSetting_defaultValue :: Lens' OptionGroupOptionSetting (Maybe Text)

The default value for the option group option.

optionGroupOptionSetting_isModifiable :: Lens' OptionGroupOptionSetting (Maybe Bool)

Boolean value where true indicates that this option group option can be changed from the default value.

optionGroupOptionSetting_settingDescription :: Lens' OptionGroupOptionSetting (Maybe Text)

The description of the option group option.

optionGroupOptionSetting_allowedValues :: Lens' OptionGroupOptionSetting (Maybe Text)

Indicates the acceptable values for the option group option.

optionGroupOptionSetting_isRequired :: Lens' OptionGroupOptionSetting (Maybe Bool)

Boolean value where true indicates that a value must be specified for this option setting of the option group option.