Network.AWS.RDS.Types.OrderableDBInstanceOption

Description

Documentation

data OrderableDBInstanceOption

Contains a list of available options for a DB instance.

This data type is used as a response element in the DescribeOrderableDBInstanceOptions action.

See: newOrderableDBInstanceOption smart constructor.

Constructors

OrderableDBInstanceOption'

Fields

minIopsPerDbInstance :: Maybe Int

Minimum total provisioned IOPS for a DB instance.

availabilityZones :: Maybe [AvailabilityZone]

A list of Availability Zones for a DB instance.

supportsIops :: Maybe Bool

Indicates whether a DB instance supports provisioned IOPS.

supportsKerberosAuthentication :: Maybe Bool

Whether a DB instance supports Kerberos Authentication.

storageType :: Maybe Text

Indicates the storage type for a DB instance.

supportsEnhancedMonitoring :: Maybe Bool

Indicates whether a DB instance supports Enhanced Monitoring at intervals from 1 to 60 seconds.

supportedEngineModes :: Maybe [Text]

A list of the supported DB engine modes.

maxStorageSize :: Maybe Int

Maximum storage size for a DB instance.

availableProcessorFeatures :: Maybe [AvailableProcessorFeature]

A list of the available processor features for the DB instance class of a DB instance.

maxIopsPerGib :: Maybe Double

Maximum provisioned IOPS per GiB for a DB instance.

supportsStorageEncryption :: Maybe Bool

Indicates whether a DB instance supports encrypted storage.

multiAZCapable :: Maybe Bool

Indicates whether a DB instance is Multi-AZ capable.

supportedActivityStreamModes :: Maybe [Text]

The list of supported modes for Database Activity Streams. Aurora PostgreSQL returns the value [sync, async]. Aurora MySQL and RDS for Oracle return [async] only. If Database Activity Streams isn't supported, the return value is an empty list.

supportsStorageAutoscaling :: Maybe Bool

Whether Amazon RDS can automatically scale storage for DB instances that use the specified DB instance class.

engineVersion :: Maybe Text

The engine version of a DB instance.

minIopsPerGib :: Maybe Double

Minimum provisioned IOPS per GiB for a DB instance.

supportsIAMDatabaseAuthentication :: Maybe Bool

Indicates whether a DB instance supports IAM database authentication.

licenseModel :: Maybe Text

The license model for a DB instance.

maxIopsPerDbInstance :: Maybe Int

Maximum total provisioned IOPS for a DB instance.

supportsGlobalDatabases :: Maybe Bool

A value that indicates whether you can use Aurora global databases with a specific combination of other DB engine attributes.

supportsPerformanceInsights :: Maybe Bool

True if a DB instance supports Performance Insights, otherwise false.

dbInstanceClass :: Maybe Text

The DB instance class for a DB instance.

engine :: Maybe Text

The engine type of a DB instance.

minStorageSize :: Maybe Int

Minimum storage size for a DB instance.

outpostCapable :: Maybe Bool

Whether a DB instance supports RDS on Outposts.

For more information about RDS on Outposts, see Amazon RDS on Amazon Web Services Outposts in the Amazon RDS User Guide.

availabilityZoneGroup :: Maybe Text

The Availability Zone group for a DB instance.

vpc :: Maybe Bool

Indicates whether a DB instance is in a VPC.

readReplicaCapable :: Maybe Bool

Indicates whether a DB instance can have a read replica.

newOrderableDBInstanceOption :: OrderableDBInstanceOption

Create a value of OrderableDBInstanceOption with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:minIopsPerDbInstance:OrderableDBInstanceOption', orderableDBInstanceOption_minIopsPerDbInstance - Minimum total provisioned IOPS for a DB instance.

$sel:availabilityZones:OrderableDBInstanceOption', orderableDBInstanceOption_availabilityZones - A list of Availability Zones for a DB instance.

$sel:supportsIops:OrderableDBInstanceOption', orderableDBInstanceOption_supportsIops - Indicates whether a DB instance supports provisioned IOPS.

$sel:supportsKerberosAuthentication:OrderableDBInstanceOption', orderableDBInstanceOption_supportsKerberosAuthentication - Whether a DB instance supports Kerberos Authentication.

$sel:storageType:OrderableDBInstanceOption', orderableDBInstanceOption_storageType - Indicates the storage type for a DB instance.

$sel:supportsEnhancedMonitoring:OrderableDBInstanceOption', orderableDBInstanceOption_supportsEnhancedMonitoring - Indicates whether a DB instance supports Enhanced Monitoring at intervals from 1 to 60 seconds.

$sel:supportedEngineModes:OrderableDBInstanceOption', orderableDBInstanceOption_supportedEngineModes - A list of the supported DB engine modes.

$sel:maxStorageSize:OrderableDBInstanceOption', orderableDBInstanceOption_maxStorageSize - Maximum storage size for a DB instance.

$sel:availableProcessorFeatures:OrderableDBInstanceOption', orderableDBInstanceOption_availableProcessorFeatures - A list of the available processor features for the DB instance class of a DB instance.

$sel:maxIopsPerGib:OrderableDBInstanceOption', orderableDBInstanceOption_maxIopsPerGib - Maximum provisioned IOPS per GiB for a DB instance.

$sel:supportsStorageEncryption:OrderableDBInstanceOption', orderableDBInstanceOption_supportsStorageEncryption - Indicates whether a DB instance supports encrypted storage.

$sel:multiAZCapable:OrderableDBInstanceOption', orderableDBInstanceOption_multiAZCapable - Indicates whether a DB instance is Multi-AZ capable.

$sel:supportedActivityStreamModes:OrderableDBInstanceOption', orderableDBInstanceOption_supportedActivityStreamModes - The list of supported modes for Database Activity Streams. Aurora PostgreSQL returns the value [sync, async]. Aurora MySQL and RDS for Oracle return [async] only. If Database Activity Streams isn't supported, the return value is an empty list.

$sel:supportsStorageAutoscaling:OrderableDBInstanceOption', orderableDBInstanceOption_supportsStorageAutoscaling - Whether Amazon RDS can automatically scale storage for DB instances that use the specified DB instance class.

$sel:engineVersion:OrderableDBInstanceOption', orderableDBInstanceOption_engineVersion - The engine version of a DB instance.

$sel:minIopsPerGib:OrderableDBInstanceOption', orderableDBInstanceOption_minIopsPerGib - Minimum provisioned IOPS per GiB for a DB instance.

$sel:supportsIAMDatabaseAuthentication:OrderableDBInstanceOption', orderableDBInstanceOption_supportsIAMDatabaseAuthentication - Indicates whether a DB instance supports IAM database authentication.

$sel:licenseModel:OrderableDBInstanceOption', orderableDBInstanceOption_licenseModel - The license model for a DB instance.

$sel:maxIopsPerDbInstance:OrderableDBInstanceOption', orderableDBInstanceOption_maxIopsPerDbInstance - Maximum total provisioned IOPS for a DB instance.

$sel:supportsGlobalDatabases:OrderableDBInstanceOption', orderableDBInstanceOption_supportsGlobalDatabases - A value that indicates whether you can use Aurora global databases with a specific combination of other DB engine attributes.

$sel:supportsPerformanceInsights:OrderableDBInstanceOption', orderableDBInstanceOption_supportsPerformanceInsights - True if a DB instance supports Performance Insights, otherwise false.

$sel:dbInstanceClass:OrderableDBInstanceOption', orderableDBInstanceOption_dbInstanceClass - The DB instance class for a DB instance.

$sel:engine:OrderableDBInstanceOption', orderableDBInstanceOption_engine - The engine type of a DB instance.

$sel:minStorageSize:OrderableDBInstanceOption', orderableDBInstanceOption_minStorageSize - Minimum storage size for a DB instance.

$sel:outpostCapable:OrderableDBInstanceOption', orderableDBInstanceOption_outpostCapable - Whether a DB instance supports RDS on Outposts.

For more information about RDS on Outposts, see Amazon RDS on Amazon Web Services Outposts in the Amazon RDS User Guide.

$sel:availabilityZoneGroup:OrderableDBInstanceOption', orderableDBInstanceOption_availabilityZoneGroup - The Availability Zone group for a DB instance.

$sel:vpc:OrderableDBInstanceOption', orderableDBInstanceOption_vpc - Indicates whether a DB instance is in a VPC.

$sel:readReplicaCapable:OrderableDBInstanceOption', orderableDBInstanceOption_readReplicaCapable - Indicates whether a DB instance can have a read replica.

orderableDBInstanceOption_minIopsPerDbInstance :: Lens' OrderableDBInstanceOption (Maybe Int)

Minimum total provisioned IOPS for a DB instance.

orderableDBInstanceOption_availabilityZones :: Lens' OrderableDBInstanceOption (Maybe [AvailabilityZone])

A list of Availability Zones for a DB instance.

orderableDBInstanceOption_supportsIops :: Lens' OrderableDBInstanceOption (Maybe Bool)

Indicates whether a DB instance supports provisioned IOPS.

orderableDBInstanceOption_supportsKerberosAuthentication :: Lens' OrderableDBInstanceOption (Maybe Bool)

Whether a DB instance supports Kerberos Authentication.

orderableDBInstanceOption_storageType :: Lens' OrderableDBInstanceOption (Maybe Text)

Indicates the storage type for a DB instance.

orderableDBInstanceOption_supportsEnhancedMonitoring :: Lens' OrderableDBInstanceOption (Maybe Bool)

Indicates whether a DB instance supports Enhanced Monitoring at intervals from 1 to 60 seconds.

orderableDBInstanceOption_supportedEngineModes :: Lens' OrderableDBInstanceOption (Maybe [Text])

A list of the supported DB engine modes.

orderableDBInstanceOption_maxStorageSize :: Lens' OrderableDBInstanceOption (Maybe Int)

Maximum storage size for a DB instance.

orderableDBInstanceOption_availableProcessorFeatures :: Lens' OrderableDBInstanceOption (Maybe [AvailableProcessorFeature])

A list of the available processor features for the DB instance class of a DB instance.

orderableDBInstanceOption_maxIopsPerGib :: Lens' OrderableDBInstanceOption (Maybe Double)

Maximum provisioned IOPS per GiB for a DB instance.

orderableDBInstanceOption_supportsStorageEncryption :: Lens' OrderableDBInstanceOption (Maybe Bool)

Indicates whether a DB instance supports encrypted storage.

orderableDBInstanceOption_multiAZCapable :: Lens' OrderableDBInstanceOption (Maybe Bool)

Indicates whether a DB instance is Multi-AZ capable.

orderableDBInstanceOption_supportedActivityStreamModes :: Lens' OrderableDBInstanceOption (Maybe [Text])

The list of supported modes for Database Activity Streams. Aurora PostgreSQL returns the value [sync, async]. Aurora MySQL and RDS for Oracle return [async] only. If Database Activity Streams isn't supported, the return value is an empty list.

orderableDBInstanceOption_supportsStorageAutoscaling :: Lens' OrderableDBInstanceOption (Maybe Bool)

Whether Amazon RDS can automatically scale storage for DB instances that use the specified DB instance class.

orderableDBInstanceOption_engineVersion :: Lens' OrderableDBInstanceOption (Maybe Text)

The engine version of a DB instance.

orderableDBInstanceOption_minIopsPerGib :: Lens' OrderableDBInstanceOption (Maybe Double)

Minimum provisioned IOPS per GiB for a DB instance.

orderableDBInstanceOption_supportsIAMDatabaseAuthentication :: Lens' OrderableDBInstanceOption (Maybe Bool)

Indicates whether a DB instance supports IAM database authentication.

orderableDBInstanceOption_licenseModel :: Lens' OrderableDBInstanceOption (Maybe Text)

The license model for a DB instance.

orderableDBInstanceOption_maxIopsPerDbInstance :: Lens' OrderableDBInstanceOption (Maybe Int)

Maximum total provisioned IOPS for a DB instance.

orderableDBInstanceOption_supportsGlobalDatabases :: Lens' OrderableDBInstanceOption (Maybe Bool)

A value that indicates whether you can use Aurora global databases with a specific combination of other DB engine attributes.

orderableDBInstanceOption_supportsPerformanceInsights :: Lens' OrderableDBInstanceOption (Maybe Bool)

True if a DB instance supports Performance Insights, otherwise false.

orderableDBInstanceOption_dbInstanceClass :: Lens' OrderableDBInstanceOption (Maybe Text)

The DB instance class for a DB instance.

orderableDBInstanceOption_engine :: Lens' OrderableDBInstanceOption (Maybe Text)

The engine type of a DB instance.

orderableDBInstanceOption_minStorageSize :: Lens' OrderableDBInstanceOption (Maybe Int)

Minimum storage size for a DB instance.

orderableDBInstanceOption_outpostCapable :: Lens' OrderableDBInstanceOption (Maybe Bool)

Whether a DB instance supports RDS on Outposts.

For more information about RDS on Outposts, see Amazon RDS on Amazon Web Services Outposts in the Amazon RDS User Guide.

orderableDBInstanceOption_availabilityZoneGroup :: Lens' OrderableDBInstanceOption (Maybe Text)

The Availability Zone group for a DB instance.

orderableDBInstanceOption_vpc :: Lens' OrderableDBInstanceOption (Maybe Bool)

Indicates whether a DB instance is in a VPC.

orderableDBInstanceOption_readReplicaCapable :: Lens' OrderableDBInstanceOption (Maybe Bool)

Indicates whether a DB instance can have a read replica.