Network.AWS.RDS.Types.PendingModifiedValues

Description

Documentation

data PendingModifiedValues

This data type is used as a response element in the ModifyDBInstance operation and contains changes that will be applied during the next maintenance window.

See: newPendingModifiedValues smart constructor.

Constructors

PendingModifiedValues'

Fields

backupRetentionPeriod :: Maybe Int

The number of days for which automated backups are retained.

cACertificateIdentifier :: Maybe Text

The identifier of the CA certificate for the DB instance.

storageType :: Maybe Text

The storage type of the DB instance.

dbSubnetGroupName :: Maybe Text

The DB subnet group for the DB instance.

masterUserPassword :: Maybe Text

The master credentials for the DB instance.

multiAZ :: Maybe Bool

A value that indicates that the Single-AZ DB instance will change to a Multi-AZ deployment.

pendingCloudwatchLogsExports :: Maybe PendingCloudwatchLogsExports

engineVersion :: Maybe Text

The database engine version.

licenseModel :: Maybe Text

The license model for the DB instance.

Valid values: license-included | bring-your-own-license | general-public-license

dbInstanceIdentifier :: Maybe Text

The database identifier for the DB instance.

processorFeatures :: Maybe [ProcessorFeature]

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

dbInstanceClass :: Maybe Text

The name of the compute and memory capacity class for the DB instance.

port :: Maybe Int

The port for the DB instance.

allocatedStorage :: Maybe Int

The allocated storage size for the DB instance specified in gibibytes (GiB).

iops :: Maybe Int

The Provisioned IOPS value for the DB instance.

iAMDatabaseAuthenticationEnabled :: Maybe Bool

Whether mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts is enabled.

newPendingModifiedValues :: PendingModifiedValues

Create a value of PendingModifiedValues with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:backupRetentionPeriod:PendingModifiedValues', pendingModifiedValues_backupRetentionPeriod - The number of days for which automated backups are retained.

$sel:cACertificateIdentifier:PendingModifiedValues', pendingModifiedValues_cACertificateIdentifier - The identifier of the CA certificate for the DB instance.

$sel:storageType:PendingModifiedValues', pendingModifiedValues_storageType - The storage type of the DB instance.

$sel:dbSubnetGroupName:PendingModifiedValues', pendingModifiedValues_dbSubnetGroupName - The DB subnet group for the DB instance.

$sel:masterUserPassword:PendingModifiedValues', pendingModifiedValues_masterUserPassword - The master credentials for the DB instance.

$sel:multiAZ:PendingModifiedValues', pendingModifiedValues_multiAZ - A value that indicates that the Single-AZ DB instance will change to a Multi-AZ deployment.

$sel:pendingCloudwatchLogsExports:PendingModifiedValues', pendingModifiedValues_pendingCloudwatchLogsExports - Undocumented member.

$sel:engineVersion:PendingModifiedValues', pendingModifiedValues_engineVersion - The database engine version.

$sel:licenseModel:PendingModifiedValues', pendingModifiedValues_licenseModel - The license model for the DB instance.

Valid values: license-included | bring-your-own-license | general-public-license

$sel:dbInstanceIdentifier:PendingModifiedValues', pendingModifiedValues_dbInstanceIdentifier - The database identifier for the DB instance.

$sel:processorFeatures:PendingModifiedValues', pendingModifiedValues_processorFeatures - The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

$sel:dbInstanceClass:PendingModifiedValues', pendingModifiedValues_dbInstanceClass - The name of the compute and memory capacity class for the DB instance.

$sel:port:PendingModifiedValues', pendingModifiedValues_port - The port for the DB instance.

$sel:allocatedStorage:PendingModifiedValues', pendingModifiedValues_allocatedStorage - The allocated storage size for the DB instance specified in gibibytes (GiB).

$sel:iops:PendingModifiedValues', pendingModifiedValues_iops - The Provisioned IOPS value for the DB instance.

$sel:iAMDatabaseAuthenticationEnabled:PendingModifiedValues', pendingModifiedValues_iAMDatabaseAuthenticationEnabled - Whether mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts is enabled.

pendingModifiedValues_backupRetentionPeriod :: Lens' PendingModifiedValues (Maybe Int)

The number of days for which automated backups are retained.

pendingModifiedValues_cACertificateIdentifier :: Lens' PendingModifiedValues (Maybe Text)

The identifier of the CA certificate for the DB instance.

pendingModifiedValues_storageType :: Lens' PendingModifiedValues (Maybe Text)

The storage type of the DB instance.

pendingModifiedValues_dbSubnetGroupName :: Lens' PendingModifiedValues (Maybe Text)

The DB subnet group for the DB instance.

pendingModifiedValues_masterUserPassword :: Lens' PendingModifiedValues (Maybe Text)

The master credentials for the DB instance.

pendingModifiedValues_multiAZ :: Lens' PendingModifiedValues (Maybe Bool)

A value that indicates that the Single-AZ DB instance will change to a Multi-AZ deployment.

pendingModifiedValues_pendingCloudwatchLogsExports :: Lens' PendingModifiedValues (Maybe PendingCloudwatchLogsExports)

Undocumented member.

pendingModifiedValues_engineVersion :: Lens' PendingModifiedValues (Maybe Text)

The database engine version.

pendingModifiedValues_licenseModel :: Lens' PendingModifiedValues (Maybe Text)

The license model for the DB instance.

Valid values: license-included | bring-your-own-license | general-public-license

pendingModifiedValues_dbInstanceIdentifier :: Lens' PendingModifiedValues (Maybe Text)

The database identifier for the DB instance.

pendingModifiedValues_processorFeatures :: Lens' PendingModifiedValues (Maybe [ProcessorFeature])

The number of CPU cores and the number of threads per core for the DB instance class of the DB instance.

pendingModifiedValues_dbInstanceClass :: Lens' PendingModifiedValues (Maybe Text)

The name of the compute and memory capacity class for the DB instance.

pendingModifiedValues_port :: Lens' PendingModifiedValues (Maybe Int)

The port for the DB instance.

pendingModifiedValues_allocatedStorage :: Lens' PendingModifiedValues (Maybe Int)

The allocated storage size for the DB instance specified in gibibytes (GiB).

pendingModifiedValues_iops :: Lens' PendingModifiedValues (Maybe Int)

The Provisioned IOPS value for the DB instance.

pendingModifiedValues_iAMDatabaseAuthenticationEnabled :: Lens' PendingModifiedValues (Maybe Bool)

Whether mapping of Amazon Web Services Identity and Access Management (IAM) accounts to database accounts is enabled.