Network.AWS.RDS.Types.ProcessorFeature

Description

Documentation

data ProcessorFeature

Contains the processor features of a DB instance class.

To specify the number of CPU cores, use the coreCount feature name for the Name parameter. To specify the number of threads per core, use the threadsPerCore feature name for the Name parameter.

You can set the processor features of the DB instance class for a DB instance when you call one of the following actions:

CreateDBInstance

ModifyDBInstance

RestoreDBInstanceFromDBSnapshot

RestoreDBInstanceFromS3

RestoreDBInstanceToPointInTime

You can view the valid processor values for a particular instance class by calling the DescribeOrderableDBInstanceOptions action and specifying the instance class for the DBInstanceClass parameter.

In addition, you can use the following actions for DB instance class processor information:

DescribeDBInstances

DescribeDBSnapshots

DescribeValidDBInstanceModifications

If you call DescribeDBInstances, ProcessorFeature returns non-null values only if the following conditions are met:

You are accessing an Oracle DB instance.

Your Oracle DB instance class supports configuring the number of CPU cores and threads per core.

The current number CPU cores and threads is set to a non-default value.

For more information, see Configuring the Processor of the DB Instance Class in the Amazon RDS User Guide.

See: newProcessorFeature smart constructor.

Constructors

ProcessorFeature'

Fields

name :: Maybe Text

The name of the processor feature. Valid names are coreCount and threadsPerCore.

value :: Maybe Text

The value of a processor feature name.

newProcessorFeature :: ProcessorFeature

Create a value of ProcessorFeature with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:name:ProcessorFeature', processorFeature_name - The name of the processor feature. Valid names are coreCount and threadsPerCore.

$sel:value:ProcessorFeature', processorFeature_value - The value of a processor feature name.

processorFeature_name :: Lens' ProcessorFeature (Maybe Text)

The name of the processor feature. Valid names are coreCount and threadsPerCore.

processorFeature_value :: Lens' ProcessorFeature (Maybe Text)

The value of a processor feature name.