Network.AWS.RDS.Types.Range

Description

Documentation

data Range

A range of integer values.

See: newRange smart constructor.

Constructors

Range'

Fields

to :: Maybe Int

The maximum value in the range.

from :: Maybe Int

The minimum value in the range.

step :: Maybe Int

The step value for the range. For example, if you have a range of 5,000 to 10,000, with a step value of 1,000, the valid values start at 5,000 and step up by 1,000. Even though 7,500 is within the range, it isn't a valid value for the range. The valid values are 5,000, 6,000, 7,000, 8,000...

newRange :: Range

Create a value of Range with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:to:Range', range_to - The maximum value in the range.

$sel:from:Range', range_from - The minimum value in the range.

$sel:step:Range', range_step - The step value for the range. For example, if you have a range of 5,000 to 10,000, with a step value of 1,000, the valid values start at 5,000 and step up by 1,000. Even though 7,500 is within the range, it isn't a valid value for the range. The valid values are 5,000, 6,000, 7,000, 8,000...

range_to :: Lens' Range (Maybe Int)

The maximum value in the range.

range_from :: Lens' Range (Maybe Int)

The minimum value in the range.

range_step :: Lens' Range (Maybe Int)

The step value for the range. For example, if you have a range of 5,000 to 10,000, with a step value of 1,000, the valid values start at 5,000 and step up by 1,000. Even though 7,500 is within the range, it isn't a valid value for the range. The valid values are 5,000, 6,000, 7,000, 8,000...