Network.AWS.RDS.Types.ReservedDBInstance

Description

Documentation

data ReservedDBInstance

This data type is used as a response element in the DescribeReservedDBInstances and PurchaseReservedDBInstancesOffering actions.

See: newReservedDBInstance smart constructor.

Constructors

ReservedDBInstance'

Fields

dbInstanceCount :: Maybe Int

The number of reserved DB instances.

state :: Maybe Text

The state of the reserved DB instance.

currencyCode :: Maybe Text

The currency code for the reserved DB instance.

startTime :: Maybe ISO8601

The time the reservation started.

productDescription :: Maybe Text

The description of the reserved DB instance.

leaseId :: Maybe Text

The unique identifier for the lease associated with the reserved DB instance.

Amazon Web Services Support might request the lease ID for an issue related to a reserved DB instance.

reservedDBInstanceId :: Maybe Text

The unique identifier for the reservation.

reservedDBInstanceArn :: Maybe Text

The Amazon Resource Name (ARN) for the reserved DB instance.

dbInstanceClass :: Maybe Text

The DB instance class for the reserved DB instance.

multiAZ :: Maybe Bool

Indicates if the reservation applies to Multi-AZ deployments.

reservedDBInstancesOfferingId :: Maybe Text

The offering identifier.

recurringCharges :: Maybe [RecurringCharge]

The recurring price charged to run this reserved DB instance.

offeringType :: Maybe Text

The offering type of this reserved DB instance.

usagePrice :: Maybe Double

The hourly price charged for this reserved DB instance.

fixedPrice :: Maybe Double

The fixed price charged for this reserved DB instance.

duration :: Maybe Int

The duration of the reservation in seconds.

newReservedDBInstance :: ReservedDBInstance

Create a value of ReservedDBInstance with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbInstanceCount:ReservedDBInstance', reservedDBInstance_dbInstanceCount - The number of reserved DB instances.

$sel:state:ReservedDBInstance', reservedDBInstance_state - The state of the reserved DB instance.

$sel:currencyCode:ReservedDBInstance', reservedDBInstance_currencyCode - The currency code for the reserved DB instance.

$sel:startTime:ReservedDBInstance', reservedDBInstance_startTime - The time the reservation started.

$sel:productDescription:ReservedDBInstance', reservedDBInstance_productDescription - The description of the reserved DB instance.

$sel:leaseId:ReservedDBInstance', reservedDBInstance_leaseId - The unique identifier for the lease associated with the reserved DB instance.

Amazon Web Services Support might request the lease ID for an issue related to a reserved DB instance.

$sel:reservedDBInstanceId:ReservedDBInstance', reservedDBInstance_reservedDBInstanceId - The unique identifier for the reservation.

$sel:reservedDBInstanceArn:ReservedDBInstance', reservedDBInstance_reservedDBInstanceArn - The Amazon Resource Name (ARN) for the reserved DB instance.

$sel:dbInstanceClass:ReservedDBInstance', reservedDBInstance_dbInstanceClass - The DB instance class for the reserved DB instance.

$sel:multiAZ:ReservedDBInstance', reservedDBInstance_multiAZ - Indicates if the reservation applies to Multi-AZ deployments.

$sel:reservedDBInstancesOfferingId:ReservedDBInstance', reservedDBInstance_reservedDBInstancesOfferingId - The offering identifier.

$sel:recurringCharges:ReservedDBInstance', reservedDBInstance_recurringCharges - The recurring price charged to run this reserved DB instance.

$sel:offeringType:ReservedDBInstance', reservedDBInstance_offeringType - The offering type of this reserved DB instance.

$sel:usagePrice:ReservedDBInstance', reservedDBInstance_usagePrice - The hourly price charged for this reserved DB instance.

$sel:fixedPrice:ReservedDBInstance', reservedDBInstance_fixedPrice - The fixed price charged for this reserved DB instance.

$sel:duration:ReservedDBInstance', reservedDBInstance_duration - The duration of the reservation in seconds.

reservedDBInstance_dbInstanceCount :: Lens' ReservedDBInstance (Maybe Int)

The number of reserved DB instances.

reservedDBInstance_state :: Lens' ReservedDBInstance (Maybe Text)

The state of the reserved DB instance.

reservedDBInstance_currencyCode :: Lens' ReservedDBInstance (Maybe Text)

The currency code for the reserved DB instance.

reservedDBInstance_startTime :: Lens' ReservedDBInstance (Maybe UTCTime)

The time the reservation started.

reservedDBInstance_productDescription :: Lens' ReservedDBInstance (Maybe Text)

The description of the reserved DB instance.

reservedDBInstance_leaseId :: Lens' ReservedDBInstance (Maybe Text)

The unique identifier for the lease associated with the reserved DB instance.

Amazon Web Services Support might request the lease ID for an issue related to a reserved DB instance.

reservedDBInstance_reservedDBInstanceId :: Lens' ReservedDBInstance (Maybe Text)

The unique identifier for the reservation.

reservedDBInstance_reservedDBInstanceArn :: Lens' ReservedDBInstance (Maybe Text)

The Amazon Resource Name (ARN) for the reserved DB instance.

reservedDBInstance_dbInstanceClass :: Lens' ReservedDBInstance (Maybe Text)

The DB instance class for the reserved DB instance.

reservedDBInstance_multiAZ :: Lens' ReservedDBInstance (Maybe Bool)

Indicates if the reservation applies to Multi-AZ deployments.

reservedDBInstance_reservedDBInstancesOfferingId :: Lens' ReservedDBInstance (Maybe Text)

The offering identifier.

reservedDBInstance_recurringCharges :: Lens' ReservedDBInstance (Maybe [RecurringCharge])

The recurring price charged to run this reserved DB instance.

reservedDBInstance_offeringType :: Lens' ReservedDBInstance (Maybe Text)

The offering type of this reserved DB instance.

reservedDBInstance_usagePrice :: Lens' ReservedDBInstance (Maybe Double)

The hourly price charged for this reserved DB instance.

reservedDBInstance_fixedPrice :: Lens' ReservedDBInstance (Maybe Double)

The fixed price charged for this reserved DB instance.

reservedDBInstance_duration :: Lens' ReservedDBInstance (Maybe Int)

The duration of the reservation in seconds.