Network.AWS.RDS.Types.ScalingConfiguration

Description

Documentation

data ScalingConfiguration

Contains the scaling configuration of an Aurora Serverless DB cluster.

For more information, see Using Amazon Aurora Serverless in the Amazon Aurora User Guide.

See: newScalingConfiguration smart constructor.

Constructors

ScalingConfiguration'

Fields

maxCapacity :: Maybe Int

The maximum capacity for an Aurora DB cluster in serverless DB engine mode.

For Aurora MySQL, valid capacity values are 1, 2, 4, 8, 16, 32, 64, 128, and 256.

For Aurora PostgreSQL, valid capacity values are 2, 4, 8, 16, 32, 64, 192, and 384.

The maximum capacity must be greater than or equal to the minimum capacity.

timeoutAction :: Maybe Text

The action to take when the timeout is reached, either ForceApplyCapacityChange or RollbackCapacityChange.

ForceApplyCapacityChange sets the capacity to the specified value as soon as possible.

RollbackCapacityChange, the default, ignores the capacity change if a scaling point isn't found in the timeout period.

If you specify ForceApplyCapacityChange, connections that prevent Aurora Serverless from finding a scaling point might be dropped.

For more information, see Autoscaling for Aurora Serverless in the Amazon Aurora User Guide.

autoPause :: Maybe Bool

A value that indicates whether to allow or disallow automatic pause for an Aurora DB cluster in serverless DB engine mode. A DB cluster can be paused only when it's idle (it has no connections).

If a DB cluster is paused for more than seven days, the DB cluster might be backed up with a snapshot. In this case, the DB cluster is restored when there is a request to connect to it.

secondsUntilAutoPause :: Maybe Int

The time, in seconds, before an Aurora DB cluster in serverless mode is paused.

Specify a value between 300 and 86,400 seconds.

secondsBeforeTimeout :: Maybe Int

The amount of time, in seconds, that Aurora Serverless tries to find a scaling point to perform seamless scaling before enforcing the timeout action. The default is 300.

Specify a value between 60 and 600 seconds.

minCapacity :: Maybe Int

The minimum capacity for an Aurora DB cluster in serverless DB engine mode.

For Aurora MySQL, valid capacity values are 1, 2, 4, 8, 16, 32, 64, 128, and 256.

For Aurora PostgreSQL, valid capacity values are 2, 4, 8, 16, 32, 64, 192, and 384.

The minimum capacity must be less than or equal to the maximum capacity.

newScalingConfiguration :: ScalingConfiguration

Create a value of ScalingConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxCapacity:ScalingConfiguration', scalingConfiguration_maxCapacity - The maximum capacity for an Aurora DB cluster in serverless DB engine mode.

For Aurora MySQL, valid capacity values are 1, 2, 4, 8, 16, 32, 64, 128, and 256.

For Aurora PostgreSQL, valid capacity values are 2, 4, 8, 16, 32, 64, 192, and 384.

The maximum capacity must be greater than or equal to the minimum capacity.

$sel:timeoutAction:ScalingConfiguration', scalingConfiguration_timeoutAction - The action to take when the timeout is reached, either ForceApplyCapacityChange or RollbackCapacityChange.

ForceApplyCapacityChange sets the capacity to the specified value as soon as possible.

RollbackCapacityChange, the default, ignores the capacity change if a scaling point isn't found in the timeout period.

If you specify ForceApplyCapacityChange, connections that prevent Aurora Serverless from finding a scaling point might be dropped.

For more information, see Autoscaling for Aurora Serverless in the Amazon Aurora User Guide.

$sel:autoPause:ScalingConfiguration', scalingConfiguration_autoPause - A value that indicates whether to allow or disallow automatic pause for an Aurora DB cluster in serverless DB engine mode. A DB cluster can be paused only when it's idle (it has no connections).

If a DB cluster is paused for more than seven days, the DB cluster might be backed up with a snapshot. In this case, the DB cluster is restored when there is a request to connect to it.

$sel:secondsUntilAutoPause:ScalingConfiguration', scalingConfiguration_secondsUntilAutoPause - The time, in seconds, before an Aurora DB cluster in serverless mode is paused.

Specify a value between 300 and 86,400 seconds.

$sel:secondsBeforeTimeout:ScalingConfiguration', scalingConfiguration_secondsBeforeTimeout - The amount of time, in seconds, that Aurora Serverless tries to find a scaling point to perform seamless scaling before enforcing the timeout action. The default is 300.

Specify a value between 60 and 600 seconds.

$sel:minCapacity:ScalingConfiguration', scalingConfiguration_minCapacity - The minimum capacity for an Aurora DB cluster in serverless DB engine mode.

For Aurora MySQL, valid capacity values are 1, 2, 4, 8, 16, 32, 64, 128, and 256.

For Aurora PostgreSQL, valid capacity values are 2, 4, 8, 16, 32, 64, 192, and 384.

The minimum capacity must be less than or equal to the maximum capacity.

scalingConfiguration_maxCapacity :: Lens' ScalingConfiguration (Maybe Int)

The maximum capacity for an Aurora DB cluster in serverless DB engine mode.

For Aurora MySQL, valid capacity values are 1, 2, 4, 8, 16, 32, 64, 128, and 256.

For Aurora PostgreSQL, valid capacity values are 2, 4, 8, 16, 32, 64, 192, and 384.

The maximum capacity must be greater than or equal to the minimum capacity.

scalingConfiguration_timeoutAction :: Lens' ScalingConfiguration (Maybe Text)

The action to take when the timeout is reached, either ForceApplyCapacityChange or RollbackCapacityChange.

ForceApplyCapacityChange sets the capacity to the specified value as soon as possible.

RollbackCapacityChange, the default, ignores the capacity change if a scaling point isn't found in the timeout period.

If you specify ForceApplyCapacityChange, connections that prevent Aurora Serverless from finding a scaling point might be dropped.

For more information, see Autoscaling for Aurora Serverless in the Amazon Aurora User Guide.

scalingConfiguration_autoPause :: Lens' ScalingConfiguration (Maybe Bool)

A value that indicates whether to allow or disallow automatic pause for an Aurora DB cluster in serverless DB engine mode. A DB cluster can be paused only when it's idle (it has no connections).

If a DB cluster is paused for more than seven days, the DB cluster might be backed up with a snapshot. In this case, the DB cluster is restored when there is a request to connect to it.

scalingConfiguration_secondsUntilAutoPause :: Lens' ScalingConfiguration (Maybe Int)

The time, in seconds, before an Aurora DB cluster in serverless mode is paused.

Specify a value between 300 and 86,400 seconds.

scalingConfiguration_secondsBeforeTimeout :: Lens' ScalingConfiguration (Maybe Int)

The amount of time, in seconds, that Aurora Serverless tries to find a scaling point to perform seamless scaling before enforcing the timeout action. The default is 300.

Specify a value between 60 and 600 seconds.

scalingConfiguration_minCapacity :: Lens' ScalingConfiguration (Maybe Int)

The minimum capacity for an Aurora DB cluster in serverless DB engine mode.

For Aurora MySQL, valid capacity values are 1, 2, 4, 8, 16, 32, 64, 128, and 256.

For Aurora PostgreSQL, valid capacity values are 2, 4, 8, 16, 32, 64, 192, and 384.

The minimum capacity must be less than or equal to the maximum capacity.