Network.AWS.RDS.Types.SourceRegion

Description

Documentation

data SourceRegion

Contains an Amazon Web Services Region name as the result of a successful call to the DescribeSourceRegions action.

See: newSourceRegion smart constructor.

Constructors

SourceRegion'

Fields

supportsDBInstanceAutomatedBackupsReplication :: Maybe Bool

Whether the source Amazon Web Services Region supports replicating automated backups to the current Amazon Web Services Region.

status :: Maybe Text

The status of the source Amazon Web Services Region.

regionName :: Maybe Text

The name of the source Amazon Web Services Region.

endpoint :: Maybe Text

The endpoint for the source Amazon Web Services Region endpoint.

newSourceRegion :: SourceRegion

Create a value of SourceRegion with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:supportsDBInstanceAutomatedBackupsReplication:SourceRegion', sourceRegion_supportsDBInstanceAutomatedBackupsReplication - Whether the source Amazon Web Services Region supports replicating automated backups to the current Amazon Web Services Region.

$sel:status:SourceRegion', sourceRegion_status - The status of the source Amazon Web Services Region.

$sel:regionName:SourceRegion', sourceRegion_regionName - The name of the source Amazon Web Services Region.

$sel:endpoint:SourceRegion', sourceRegion_endpoint - The endpoint for the source Amazon Web Services Region endpoint.

sourceRegion_supportsDBInstanceAutomatedBackupsReplication :: Lens' SourceRegion (Maybe Bool)

Whether the source Amazon Web Services Region supports replicating automated backups to the current Amazon Web Services Region.

sourceRegion_status :: Lens' SourceRegion (Maybe Text)

The status of the source Amazon Web Services Region.

sourceRegion_regionName :: Lens' SourceRegion (Maybe Text)

The name of the source Amazon Web Services Region.

sourceRegion_endpoint :: Lens' SourceRegion (Maybe Text)

The endpoint for the source Amazon Web Services Region endpoint.