Network.AWS.RDS.Types.UpgradeTarget

Description

Documentation

data UpgradeTarget

The version of the database engine that a DB instance can be upgraded to.

See: newUpgradeTarget smart constructor.

Constructors

UpgradeTarget'

Fields

engineVersion :: Maybe Text

The version number of the upgrade target database engine.

supportedEngineModes :: Maybe [Text]

A list of the supported DB engine modes for the target engine version.

isMajorVersionUpgrade :: Maybe Bool

A value that indicates whether upgrading to the target version requires upgrading the major version of the database engine.

engine :: Maybe Text

The name of the upgrade target database engine.

supportsGlobalDatabases :: Maybe Bool

A value that indicates whether you can use Aurora global databases with the target engine version.

autoUpgrade :: Maybe Bool

A value that indicates whether the target version is applied to any source DB instances that have AutoMinorVersionUpgrade set to true.

supportsParallelQuery :: Maybe Bool

A value that indicates whether you can use Aurora parallel query with the target engine version.

description :: Maybe Text

The version of the database engine that a DB instance can be upgraded to.

newUpgradeTarget :: UpgradeTarget

Create a value of UpgradeTarget with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:engineVersion:UpgradeTarget', upgradeTarget_engineVersion - The version number of the upgrade target database engine.

$sel:supportedEngineModes:UpgradeTarget', upgradeTarget_supportedEngineModes - A list of the supported DB engine modes for the target engine version.

$sel:isMajorVersionUpgrade:UpgradeTarget', upgradeTarget_isMajorVersionUpgrade - A value that indicates whether upgrading to the target version requires upgrading the major version of the database engine.

$sel:engine:UpgradeTarget', upgradeTarget_engine - The name of the upgrade target database engine.

$sel:supportsGlobalDatabases:UpgradeTarget', upgradeTarget_supportsGlobalDatabases - A value that indicates whether you can use Aurora global databases with the target engine version.

$sel:autoUpgrade:UpgradeTarget', upgradeTarget_autoUpgrade - A value that indicates whether the target version is applied to any source DB instances that have AutoMinorVersionUpgrade set to true.

$sel:supportsParallelQuery:UpgradeTarget', upgradeTarget_supportsParallelQuery - A value that indicates whether you can use Aurora parallel query with the target engine version.

$sel:description:UpgradeTarget', upgradeTarget_description - The version of the database engine that a DB instance can be upgraded to.

upgradeTarget_engineVersion :: Lens' UpgradeTarget (Maybe Text)

The version number of the upgrade target database engine.

upgradeTarget_supportedEngineModes :: Lens' UpgradeTarget (Maybe [Text])

A list of the supported DB engine modes for the target engine version.

upgradeTarget_isMajorVersionUpgrade :: Lens' UpgradeTarget (Maybe Bool)

A value that indicates whether upgrading to the target version requires upgrading the major version of the database engine.

upgradeTarget_engine :: Lens' UpgradeTarget (Maybe Text)

The name of the upgrade target database engine.

upgradeTarget_supportsGlobalDatabases :: Lens' UpgradeTarget (Maybe Bool)

A value that indicates whether you can use Aurora global databases with the target engine version.

upgradeTarget_autoUpgrade :: Lens' UpgradeTarget (Maybe Bool)

A value that indicates whether the target version is applied to any source DB instances that have AutoMinorVersionUpgrade set to true.

upgradeTarget_supportsParallelQuery :: Lens' UpgradeTarget (Maybe Bool)

A value that indicates whether you can use Aurora parallel query with the target engine version.

upgradeTarget_description :: Lens' UpgradeTarget (Maybe Text)

The version of the database engine that a DB instance can be upgraded to.