Network.AWS.RDS.Types.UserAuthConfigInfo

Description

Documentation

data UserAuthConfigInfo

Returns the details of authentication used by a proxy to log in as a specific database user.

See: newUserAuthConfigInfo smart constructor.

Constructors

UserAuthConfigInfo'

Fields

secretArn :: Maybe Text

The Amazon Resource Name (ARN) representing the secret that the proxy uses to authenticate to the RDS DB instance or Aurora DB cluster. These secrets are stored within Amazon Secrets Manager.

authScheme :: Maybe AuthScheme

The type of authentication that the proxy uses for connections from the proxy to the underlying database.

iAMAuth :: Maybe IAMAuthMode

Whether to require or disallow Amazon Web Services Identity and Access Management (IAM) authentication for connections to the proxy.

userName :: Maybe Text

The name of the database user to which the proxy connects.

description :: Maybe Text

A user-specified description about the authentication used by a proxy to log in as a specific database user.

newUserAuthConfigInfo :: UserAuthConfigInfo

Create a value of UserAuthConfigInfo with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:secretArn:UserAuthConfigInfo', userAuthConfigInfo_secretArn - The Amazon Resource Name (ARN) representing the secret that the proxy uses to authenticate to the RDS DB instance or Aurora DB cluster. These secrets are stored within Amazon Secrets Manager.

$sel:authScheme:UserAuthConfigInfo', userAuthConfigInfo_authScheme - The type of authentication that the proxy uses for connections from the proxy to the underlying database.

$sel:iAMAuth:UserAuthConfigInfo', userAuthConfigInfo_iAMAuth - Whether to require or disallow Amazon Web Services Identity and Access Management (IAM) authentication for connections to the proxy.

$sel:userName:UserAuthConfigInfo', userAuthConfigInfo_userName - The name of the database user to which the proxy connects.

$sel:description:UserAuthConfigInfo', userAuthConfigInfo_description - A user-specified description about the authentication used by a proxy to log in as a specific database user.

userAuthConfigInfo_secretArn :: Lens' UserAuthConfigInfo (Maybe Text)

The Amazon Resource Name (ARN) representing the secret that the proxy uses to authenticate to the RDS DB instance or Aurora DB cluster. These secrets are stored within Amazon Secrets Manager.

userAuthConfigInfo_authScheme :: Lens' UserAuthConfigInfo (Maybe AuthScheme)

The type of authentication that the proxy uses for connections from the proxy to the underlying database.

userAuthConfigInfo_iAMAuth :: Lens' UserAuthConfigInfo (Maybe IAMAuthMode)

Whether to require or disallow Amazon Web Services Identity and Access Management (IAM) authentication for connections to the proxy.

userAuthConfigInfo_userName :: Lens' UserAuthConfigInfo (Maybe Text)

The name of the database user to which the proxy connects.

userAuthConfigInfo_description :: Lens' UserAuthConfigInfo (Maybe Text)

A user-specified description about the authentication used by a proxy to log in as a specific database user.