Network.AWS.RDS.Types.VpnDetails

Description

Documentation

data VpnDetails

Information about the virtual private network (VPN) between the VMware vSphere cluster and the Amazon Web Services website.

For more information about RDS on VMware, see the RDS on VMware User Guide.

See: newVpnDetails smart constructor.

Constructors

VpnDetails'

Fields

vpnId :: Maybe Text

The ID of the VPN.

vpnTunnelOriginatorIP :: Maybe Text

The IP address of network traffic from your on-premises data center. A custom AZ receives the network traffic.

vpnName :: Maybe Text

The name of the VPN.

vpnState :: Maybe Text

The state of the VPN.

vpnGatewayIp :: Maybe Text

The IP address of network traffic from Amazon Web Services to your on-premises data center.

vpnPSK :: Maybe (Sensitive Text)

The preshared key (PSK) for the VPN.

newVpnDetails :: VpnDetails

Create a value of VpnDetails with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:vpnId:VpnDetails', vpnDetails_vpnId - The ID of the VPN.

$sel:vpnTunnelOriginatorIP:VpnDetails', vpnDetails_vpnTunnelOriginatorIP - The IP address of network traffic from your on-premises data center. A custom AZ receives the network traffic.

$sel:vpnName:VpnDetails', vpnDetails_vpnName - The name of the VPN.

$sel:vpnState:VpnDetails', vpnDetails_vpnState - The state of the VPN.

$sel:vpnGatewayIp:VpnDetails', vpnDetails_vpnGatewayIp - The IP address of network traffic from Amazon Web Services to your on-premises data center.

$sel:vpnPSK:VpnDetails', vpnDetails_vpnPSK - The preshared key (PSK) for the VPN.

vpnDetails_vpnId :: Lens' VpnDetails (Maybe Text)

The ID of the VPN.

vpnDetails_vpnTunnelOriginatorIP :: Lens' VpnDetails (Maybe Text)

The IP address of network traffic from your on-premises data center. A custom AZ receives the network traffic.

vpnDetails_vpnName :: Lens' VpnDetails (Maybe Text)

The name of the VPN.

vpnDetails_vpnState :: Lens' VpnDetails (Maybe Text)

The state of the VPN.

vpnDetails_vpnGatewayIp :: Lens' VpnDetails (Maybe Text)

The IP address of network traffic from Amazon Web Services to your on-premises data center.

vpnDetails_vpnPSK :: Lens' VpnDetails (Maybe Text)

The preshared key (PSK) for the VPN.