Network.AWS.RDS.V2013_09_09.CreateDBParameterGroup

Description

Creates a new DB parameter group. A DB parameter group is initially created with the default parameters for the database engine used by the DB instance. To provide custom values for any of the parameters, you must modify the group after creating it using ModifyDBParameterGroup. Once you've created a DB parameter group, you need to associate it with your DB instance using ModifyDBInstance. When you associate a new DB parameter group with a running DB instance, you need to reboot the DB Instance for the new DB parameter group and associated settings to take effect. https:/rds.amazonaws.com ?Action=CreateDBParameterGroup &DBParameterGroupName=mydbparametergroup3 &DBParameterGroupFamily=MySQL5.1 &Version=2013-05-15 &Description=My%20new%20DBParameterGroup &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2011-05-11T18%3A09%3A29.793Z &AWSAccessKeyId= &Signature= mysql5.1 My new DBParameterGroup mydbparametergroup3 0b447b66-bf36-11de-a88b-7b5b3d23b3a7.

Documentation

createDBParameterGroup

Arguments

:: Text
_cdbpgmDBParameterGroupFamily

-> Text
_cdbpgmDBParameterGroupName

-> Text
_cdbpgmDescription

-> CreateDBParameterGroup

Minimum specification for a CreateDBParameterGroup request.

data CreateDBParameterGroup

Constructors

CreateDBParameterGroup

Fields

_cdbpgmDBParameterGroupFamily :: Text

The DB parameter group family name. A DB parameter group can be associated with one and only one DB parameter group family, and can be applied only to a DB instance running a database engine and engine version compatible with that DB parameter group family.

_cdbpgmDBParameterGroupName :: Text

The name of the DB parameter group. Constraints: Must be 1 to 255 alphanumeric characters First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens This value is stored as a lower-case string.

_cdbpgmDescription :: Text

The description for the DB parameter group.

_cdbpgmTags :: [Tag]

A list of tags.

cdbpgmTags :: Lens' CreateDBParameterGroup [Tag]

cdbpgmDescription :: Lens' CreateDBParameterGroup Text

cdbpgmDBParameterGroupName :: Lens' CreateDBParameterGroup Text

cdbpgmDBParameterGroupFamily :: Lens' CreateDBParameterGroup Text

data CreateDBParameterGroupResponse

Constructors

CreateDBParameterGroupResponse

Fields

_dbpgwDBParameterGroup :: Maybe DBParameterGroup

Contains the result of a successful invocation of the CreateDBParameterGroup action. This data type is used as a request parameter in the DeleteDBParameterGroup action, and as a response element in the DescribeDBParameterGroups action.

dbpgwDBParameterGroup :: Iso' CreateDBParameterGroupResponse (Maybe DBParameterGroup)