Network.AWS.RDS.V2013_09_09.CreateDBSecurityGroup

Description

Creates a new DB security group. DB security groups control access to a DB instance. https:/rds.amazonaws.com ?Action=CreateDBSecurityGroup &DBSecurityGroupName=mydbsecuritygroup &DBSecurityGroupDescription=My%20new%20DBSecurityGroup &EC2VpcId=vpc-1a2b3c4d &Version=2013-05-15 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2011-02-15T18%3A14%3A49.482Z &AWSAccessKeyId= &Signature= My new DBSecurityGroup 565419523791 mydbsecuritygroup vpc-1a2b3c4d ed662948-a57b-11df-9e38-7ffab86c801f.

Documentation

createDBSecurityGroup

Arguments

:: Text
_cdbsgmDBSecurityGroupName

-> Text
_cdbsgmDBSecurityGroupDescription

-> CreateDBSecurityGroup

Minimum specification for a CreateDBSecurityGroup request.

data CreateDBSecurityGroup

Constructors

CreateDBSecurityGroup

Fields

_cdbsgmDBSecurityGroupName :: Text

The name for the DB security group. This value is stored as a lowercase string. Constraints: Must be 1 to 255 alphanumeric characters First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens Must not be Default May not contain spaces Example: mysecuritygroup.

_cdbsgmDBSecurityGroupDescription :: Text

The description for the DB security group.

_cdbsgmTags :: [Tag]

A list of tags.

cdbsgmTags :: Lens' CreateDBSecurityGroup [Tag]

cdbsgmDBSecurityGroupName :: Lens' CreateDBSecurityGroup Text

cdbsgmDBSecurityGroupDescription :: Lens' CreateDBSecurityGroup Text

data CreateDBSecurityGroupResponse

Constructors

CreateDBSecurityGroupResponse

Fields

_dbsgzDBSecurityGroup :: Maybe DBSecurityGroup

Contains the result of a successful invocation of the following actions: DescribeDBSecurityGroups AuthorizeDBSecurityGroupIngress CreateDBSecurityGroup RevokeDBSecurityGroupIngress This data type is used as a response element in the DescribeDBSecurityGroups action.

dbsgzDBSecurityGroup :: Iso' CreateDBSecurityGroupResponse (Maybe DBSecurityGroup)