Network.AWS.RDS.V2013_09_09.CreateDBSubnetGroup

Description

Creates a new DB subnet group. DB subnet groups must contain at least one subnet in at least two AZs in the region. https:/rds.amazonaws.com ?Action=CreateDBSubnetGroup &DBSubnetGroupName=mydbsubnetgroup &DBSubnetGroupDescription=My%20new%20DBSubnetGroup &Version=2013-05-15 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2011-02-15T18%3A14%3A49.482Z &AWSAccessKeyId= &Signature= 990524496922 Complete My new DBSubnetGroup mydbsubnetgroup Active subnet-7c5b4115 us-east-1c Active subnet-7b5b4112 us-east-1b Active subnet-3ea6bd57 us-east-1d ed662948-a57b-11df-9e38-7ffab86c801f.

Documentation

createDBSubnetGroup

Arguments

:: Text
_cdbsgnDBSubnetGroupName

-> Text
_cdbsgnDBSubnetGroupDescription

-> [Text]
_cdbsgnSubnetIds

-> CreateDBSubnetGroup

Minimum specification for a CreateDBSubnetGroup request.

data CreateDBSubnetGroup

Constructors

CreateDBSubnetGroup

Fields

_cdbsgnDBSubnetGroupName :: Text

The name for the DB subnet group. This value is stored as a lowercase string. Constraints: Must contain no more than 255 alphanumeric characters or hyphens. Must not be Default. Example: mySubnetgroup.

_cdbsgnDBSubnetGroupDescription :: Text

The description for the DB subnet group.

_cdbsgnSubnetIds :: [Text]

The EC2 Subnet IDs for the DB subnet group.

_cdbsgnTags :: [Tag]

A list of tags.

cdbsgnTags :: Lens' CreateDBSubnetGroup [Tag]

cdbsgnSubnetIds :: Lens' CreateDBSubnetGroup [Text]

cdbsgnDBSubnetGroupName :: Lens' CreateDBSubnetGroup Text

cdbsgnDBSubnetGroupDescription :: Lens' CreateDBSubnetGroup Text

data CreateDBSubnetGroupResponse

Constructors

CreateDBSubnetGroupResponse

Fields

_dbsgdrDBSubnetGroup :: Maybe DBSubnetGroup

Contains the result of a successful invocation of the following actions: CreateDBSubnetGroup ModifyDBSubnetGroup DescribeDBSubnetGroups DeleteDBSubnetGroup This data type is used as a response element in the DescribeDBSubnetGroups action.

dbsgdrDBSubnetGroup :: Iso' CreateDBSubnetGroupResponse (Maybe DBSubnetGroup)