Network.AWS.RDS.V2013_09_09.DeleteDBParameterGroup

Description

Deletes a specified DBParameterGroup. The DBParameterGroup cannot be associated with any RDS instances to be deleted. The specified DB parameter group cannot be associated with any DB instances. https:/rds.amazonaws.com ?Action=DeleteDBParameterGroup &DBParameterGroupName=mydbparametergroup &Version=2013-05-15 &SignatureVersion=2&SignatureMethod=HmacSHA256 &Timestamp=2011-05-11T18%3A47%3A08.851Z &AWSAccessKeyId= &Signature= 4dc38be9-bf3b-11de-a88b-7b5b3d23b3a7.

Documentation

data DeleteDBParameterGroup

Constructors

DeleteDBParameterGroup

Fields

_ddbpgnDBParameterGroupName :: Text

The name of the DB parameter group. Constraints: Must be the name of an existing DB parameter group You cannot delete a default DB parameter group Cannot be associated with any DB instances.

ddbpgnDBParameterGroupName :: Iso' DeleteDBParameterGroup Text

data DeleteDBParameterGroupResponse

Constructors

DeleteDBParameterGroupResponse