Network.AWS.RDS.V2013_09_09.DeleteDBSubnetGroup

Description

Deletes a DB subnet group. The specified database subnet group must not be associated with any DB instances. https:/rds.amazonaws.com ?Action=DeleteDBSubnetGroup &DBSubnetGroupName=mysubnetgroup &Version=2013-05-15 &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2011-02-15T17%3A48%3A21.746Z &AWSAccessKeyId= &Signature= 5d013245-4172-11df-8520-e7e1e602a915.

Documentation

data DeleteDBSubnetGroup

Constructors

DeleteDBSubnetGroup

Fields

_ddbsgnDBSubnetGroupName :: Text

The name of the database subnet group to delete. You cannot delete the default subnet group. Constraints: Must be 1 to 255 alphanumeric characters First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens.

ddbsgnDBSubnetGroupName :: Iso' DeleteDBSubnetGroup Text

data DeleteDBSubnetGroupResponse

Constructors

DeleteDBSubnetGroupResponse