Network.AWS.RDS.V2013_09_09.DescribeDBLogFiles

Description

Returns a list of DB log files for the DB instance. https:/rds.amazonaws.com ?DBInstanceIdentifier=rrak-mysql &MaxRecords=100 &Version=2013-02-12 &Action=DescribeDBLogFiles &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20130327T173621Z &X-Amz-Algorithm=AWS4-HMAC-SHA256 &X-Amz-Date=20130327T173621Z &X-Amz-SignedHeaders=Host &X-Amz-Expires=20130327T173621Z &X-Amz-Credential= &X-Amz-Signature= 1364403600000 error/mysql-error-running.log 0 1364338800000 error/mysql-error-running.log.0 0 1364342400000 error/mysql-error-running.log.1 0 1364346000000 error/mysql-error-running.log.2 0 1364349600000 errormysql-error-running.log.3 0 1364405700000 errormysql-error.log 0 d70fb3b3-9704-11e2-a0db-871552e0ef19.

Documentation

describeDBLogFiles

Arguments

:: Text
_ddblfmDBInstanceIdentifier

-> DescribeDBLogFiles

Minimum specification for a DescribeDBLogFiles request.

data DescribeDBLogFiles

Constructors

DescribeDBLogFiles

Fields

_ddblfmDBInstanceIdentifier :: Text

The customer-assigned name of the DB instance that contains the log files you want to list. Constraints: Must contain from 1 to 63 alphanumeric characters or hyphens First character must be a letter Cannot end with a hyphen or contain two consecutive hyphens.

_ddblfmMaxRecords :: Maybe Integer

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved.

_ddblfmFileSize :: Maybe Integer

Filters the available log files for files larger than the specified size.

_ddblfmFileLastWritten :: Maybe Integer

Filters the available log files for files written since the specified date, in POSIX timestamp format.

_ddblfmFilenameContains :: Maybe Text

Filters the available log files for log file names that contain the specified string.

_ddblfmMarker :: Maybe Text

The pagination token provided in the previous request. If this parameter is specified the response includes only records beyond the marker, up to MaxRecords.

ddblfmMaxRecords :: Lens' DescribeDBLogFiles (Maybe Integer)

ddblfmMarker :: Lens' DescribeDBLogFiles (Maybe Text)

ddblfmFilenameContains :: Lens' DescribeDBLogFiles (Maybe Text)

ddblfmFileSize :: Lens' DescribeDBLogFiles (Maybe Integer)

ddblfmFileLastWritten :: Lens' DescribeDBLogFiles (Maybe Integer)

ddblfmDBInstanceIdentifier :: Lens' DescribeDBLogFiles Text

data DescribeDBLogFilesResponse

Constructors

DescribeDBLogFilesResponse

Fields

_ddblfrDescribeDBLogFiles :: [DescribeDBLogFilesDetails]

The DB log files returned.

_ddblfrMarker :: Maybe Text

An optional paging token.

ddblfrMarker :: Lens' DescribeDBLogFilesResponse (Maybe Text)

ddblfrDescribeDBLogFiles :: Lens' DescribeDBLogFilesResponse [DescribeDBLogFilesDetails]