Network.AWS.RDS.V2013_09_09.DescribeEventCategories

Description

Displays a list of categories for all event source types, or, if specified, for a specified source type. You can see a list of the event categories and source types in the Events topic in the Amazon RDS User Guide. https:/rds.us-east-1.amazonaws.com ?Action=DescribeEventCategories &SourceType=db-instance &Version=2013-01-10 &SignatureVersion=4 &SignatureMethod=HmacSHA256 &Timestamp=20130128T013452Z &AWSAccessKeyId= &Signature= db-instance failover low storage maintenance recovery restoration deletion configuration change failover availability creation backup notification ea3bf54b-68ea-11e2-bd13-a92da73b3119.

Documentation

describeEventCategories :: DescribeEventCategories

Minimum specification for a DescribeEventCategories request.

data DescribeEventCategories

Constructors

DescribeEventCategories

Fields

_decmSourceType :: Maybe Text

The type of source that will be generating the events. Valid values: db-instance | db-parameter-group | db-security-group | db-snapshot.

decmSourceType :: Iso' DescribeEventCategories (Maybe Text)

data DescribeEventCategoriesResponse

Constructors

DescribeEventCategoriesResponse

Fields

_ecvEventCategoriesMapList :: [EventCategoriesMap]

A list of EventCategoriesMap data types.

ecvEventCategoriesMapList :: Iso' DescribeEventCategoriesResponse [EventCategoriesMap]