Network.AWS.RDS.V2013_09_09.DescribeOptionGroupOptions

Description

Describes all available options. https:/rds.amazonaws.com ?Action=DescribeOptionGroupOptions &EngineName=oracle-se1 &MajorEngineVersion=11.2 11.2 true Oracle Enterprise Manager 1158 OEM oracle-se1 0.2.v3 false false d9c8f6a1-84c7-11e1-a264-0b23c28bc344.

Documentation

describeOptionGroupOptions

Arguments

:: Text
_dogomEngineName

-> DescribeOptionGroupOptions

Minimum specification for a DescribeOptionGroupOptions request.

data DescribeOptionGroupOptions

Constructors

DescribeOptionGroupOptions

Fields

_dogomEngineName :: Text

A required parameter. Options available for the given Engine name will be described.

_dogomMaxRecords :: Maybe Integer

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20, maximum 100.

_dogomMajorEngineVersion :: Maybe Text

If specified, filters the results to include only options for the specified major engine version.

_dogomMarker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

dogomMaxRecords :: Lens' DescribeOptionGroupOptions (Maybe Integer)

dogomMarker :: Lens' DescribeOptionGroupOptions (Maybe Text)

dogomMajorEngineVersion :: Lens' DescribeOptionGroupOptions (Maybe Text)

dogomEngineName :: Lens' DescribeOptionGroupOptions Text

data DescribeOptionGroupOptionsResponse

Constructors

DescribeOptionGroupOptionsResponse

Fields

_ogomOptionGroupOptions :: [OptionGroupOption]

List of available option group options.

_ogomMarker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

ogomOptionGroupOptions :: Lens' DescribeOptionGroupOptionsResponse [OptionGroupOption]

ogomMarker :: Lens' DescribeOptionGroupOptionsResponse (Maybe Text)