Network.AWS.RDS.V2013_09_09.DescribeOptionGroups

Description

Describes the available option groups. https:/rds.amazonaws.com ?Action=DescribeOptionGroups &OptionGroupName=myoptiongroup &MaxRecords=100 11.2 myoptiongroup oracle-se1 Test option group 6088823d-84c8-11e1-a264-0b23c28bc344 https:/rds.amazonaws.com ?Action=DescribeOptionGroups &MaxRecords=100 11.2 myoptiongroup oracle-se1 Test option group 11.2 default:oracle-se1-11-2 oracle-se1 Default option group. e4b234d9-84d5-11e1-87a6-71059839a52b.

Documentation

describeOptionGroups :: DescribeOptionGroups

Minimum specification for a DescribeOptionGroups request.

data DescribeOptionGroups

Constructors

DescribeOptionGroups

Fields

_dogmMaxRecords :: Maybe Integer

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20, maximum 100.

_dogmEngineName :: Maybe Text

Filters the list of option groups to only include groups associated with a specific database engine.

_dogmMajorEngineVersion :: Maybe Text

Filters the list of option groups to only include groups associated with a specific database engine version. If specified, then EngineName must also be specified.

_dogmMarker :: Maybe Text

An optional pagination token provided by a previous DescribeOptionGroups request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

_dogmOptionGroupName :: Maybe Text

The name of the option group to describe. Cannot be supplied together with EngineName or MajorEngineVersion.

dogmOptionGroupName :: Lens' DescribeOptionGroups (Maybe Text)

dogmMaxRecords :: Lens' DescribeOptionGroups (Maybe Integer)

dogmMarker :: Lens' DescribeOptionGroups (Maybe Text)

dogmMajorEngineVersion :: Lens' DescribeOptionGroups (Maybe Text)

dogmEngineName :: Lens' DescribeOptionGroups (Maybe Text)

data DescribeOptionGroupsResponse

Constructors

DescribeOptionGroupsResponse

Fields

_ogOptionGroupsList :: [OptionGroup]

List of option groups.

_ogMarker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

ogOptionGroupsList :: Lens' DescribeOptionGroupsResponse [OptionGroup]

ogMarker :: Lens' DescribeOptionGroupsResponse (Maybe Text)