Network.AWS.RDS.V2013_09_09.DescribeOrderableDBInstanceOptions

Description

Returns a list of orderable DB instance options for the specified engine. https:/rds.amazonaws.com ?Action=DescribeOrderableDBInstanceOptions &Engine=mysql &MaxRecords=100 &Version=2013-05-15 &Timestamp=2011-05-23T07%3A49%3A17.749Z &SignatureVersion=2 &SignatureMethod=HmacSHA256 &AWSAccessKeyId= &Signature= true mysql general-public-license true 5.1.45 db.m1.large us-east-1a yes us-east-1b no us-east-1d yes true mysql general-public-license true 5.1.45 db.m1.small us-east-1a yes us-east-1b yes us-east-1d yes true mysql general-public-license true 5.1.45 db.m1.xlarge us-east-1a yes us-east-1b yes us-east-1d yes true mysql general-public-license true 5.1.45 db.m2.2xlarge us-east-1a yes us-east-1b yes us-east-1d yes true mysql general-public-license true 5.1.45 db.m2.4xlarge us-east-1a yes us-east-1b no us-east-1d no 2a0406d7-8511-11e0-90aa-eb648410240d.

Documentation

describeOrderableDBInstanceOptions

Arguments

:: Text
_dodbiomEngine

-> DescribeOrderableDBInstanceOptions

Minimum specification for a DescribeOrderableDBInstanceOptions request.

data DescribeOrderableDBInstanceOptions

Constructors

DescribeOrderableDBInstanceOptions

Fields

_dodbiomEngine :: Text

The name of the engine to retrieve DB instance options for.

_dodbiomVpc :: Maybe Bool

The VPC filter value. Specify this parameter to show only the available VPC or non-VPC offerings.

_dodbiomMaxRecords :: Maybe Integer

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: minimum 20, maximum 100.

_dodbiomEngineVersion :: Maybe Text

The engine version filter value. Specify this parameter to show only the available offerings matching the specified engine version.

_dodbiomDBInstanceClass :: Maybe Text

The DB instance class filter value. Specify this parameter to show only the available offerings matching the specified DB instance class.

_dodbiomLicenseModel :: Maybe Text

The license model filter value. Specify this parameter to show only the available offerings matching the specified license model.

_dodbiomMarker :: Maybe Text

An optional pagination token provided by a previous DescribeOrderableDBInstanceOptions request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dodbiomVpc :: Lens' DescribeOrderableDBInstanceOptions (Maybe Bool)

dodbiomMaxRecords :: Lens' DescribeOrderableDBInstanceOptions (Maybe Integer)

dodbiomMarker :: Lens' DescribeOrderableDBInstanceOptions (Maybe Text)

dodbiomLicenseModel :: Lens' DescribeOrderableDBInstanceOptions (Maybe Text)

dodbiomEngineVersion :: Lens' DescribeOrderableDBInstanceOptions (Maybe Text)

dodbiomEngine :: Lens' DescribeOrderableDBInstanceOptions Text

dodbiomDBInstanceClass :: Lens' DescribeOrderableDBInstanceOptions (Maybe Text)

data DescribeOrderableDBInstanceOptionsResponse

Constructors

DescribeOrderableDBInstanceOptionsResponse

Fields

_odbiomOrderableDBInstanceOptions :: [OrderableDBInstanceOption]

An OrderableDBInstanceOption structure containing information about orderable options for the DB instance.

_odbiomMarker :: Maybe Text

An optional pagination token provided by a previous OrderableDBInstanceOptions request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

odbiomOrderableDBInstanceOptions :: Lens' DescribeOrderableDBInstanceOptionsResponse [OrderableDBInstanceOption]

odbiomMarker :: Lens' DescribeOrderableDBInstanceOptionsResponse (Maybe Text)