Network.AWS.RDS.V2013_09_09.DescribeReservedDBInstances

Description

Returns information about reserved DB instances for this account, or about a specified reserved DB instance. https:/rds.amazonaws.com ?Action=DescribeReservedDBInstances &ReservedDBInstanceId=customerSpecifiedID &SignatureVersion=2 &SignatureMethod=HmacSHA256 &Timestamp=2012-12-18T18%3A31%3A36.118Z &AWSAccessKeyId= &Signature= Medium Utilization USD mysql 649fd0c8-cf6d-47a0-bfa6-060f8e75e95f false active myreservationid 1 2010-12-15T00:25:14.131Z 31536000 227.5 0.046 db.m1.small c695119b-2961-11e1-bd06-6fe008f046c3.

Documentation

describeReservedDBInstances :: DescribeReservedDBInstances

Minimum specification for a DescribeReservedDBInstances request.

data DescribeReservedDBInstances

Constructors

DescribeReservedDBInstances

Fields

_drdbimMultiAZ :: Maybe Bool

The Multi-AZ filter value. Specify this parameter to show only those reservations matching the specified Multi-AZ parameter.

_drdbimMaxRecords :: Maybe Integer

The maximum number of records to include in the response. If more than the MaxRecords value is available, a pagination token called a marker is included in the response so that the following results can be retrieved. Default: 100 Constraints: minimum 20, maximum 100.

_drdbimProductDescription :: Maybe Text

The product description filter value. Specify this parameter to show only those reservations matching the specified product description.

_drdbimReservedDBInstanceId :: Maybe Text

The reserved DB instance identifier filter value. Specify this parameter to show only the reservation that matches the specified reservation ID.

_drdbimDBInstanceClass :: Maybe Text

The DB instance class filter value. Specify this parameter to show only those reservations matching the specified DB instances class.

_drdbimMarker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

_drdbimReservedDBInstancesOfferingId :: Maybe Text

The offering identifier filter value. Specify this parameter to show only purchased reservations matching the specified offering identifier.

_drdbimOfferingType :: Maybe Text

The offering type filter value. Specify this parameter to show only the available offerings matching the specified offering type. Valid Values: "Light Utilization" | "Medium Utilization" | "Heavy Utilization".

_drdbimDuration :: Maybe Text

The duration filter value, specified in years or seconds. Specify this parameter to show only reservations for this duration. Valid Values: 1 | 3 | 31536000 | 94608000.

drdbimReservedDBInstancesOfferingId :: Lens' DescribeReservedDBInstances (Maybe Text)

drdbimReservedDBInstanceId :: Lens' DescribeReservedDBInstances (Maybe Text)

drdbimProductDescription :: Lens' DescribeReservedDBInstances (Maybe Text)

drdbimOfferingType :: Lens' DescribeReservedDBInstances (Maybe Text)

drdbimMultiAZ :: Lens' DescribeReservedDBInstances (Maybe Bool)

drdbimMaxRecords :: Lens' DescribeReservedDBInstances (Maybe Integer)

drdbimMarker :: Lens' DescribeReservedDBInstances (Maybe Text)

drdbimDuration :: Lens' DescribeReservedDBInstances (Maybe Text)

drdbimDBInstanceClass :: Lens' DescribeReservedDBInstances (Maybe Text)

data DescribeReservedDBInstancesResponse

Constructors

DescribeReservedDBInstancesResponse

Fields

_rdbimReservedDBInstances :: [ReservedDBInstance]

A list of reserved DB instances.

_rdbimMarker :: Maybe Text

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords.

rdbimReservedDBInstances :: Lens' DescribeReservedDBInstancesResponse [ReservedDBInstance]

rdbimMarker :: Lens' DescribeReservedDBInstancesResponse (Maybe Text)