Network.AWS.RDSData

Description

Derived from API version 2018-08-01 of the AWS service descriptions, licensed under Apache 2.0.

Amazon RDS Data Service

Amazon RDS provides an HTTP endpoint to run SQL statements on an Amazon Aurora Serverless DB cluster. To run these statements, you work with the Data Service API.

For more information about the Data Service API, see Using the Data API for Aurora Serverless in the Amazon Aurora User Guide.

Service Configuration

defaultService :: Service

API version 2018-08-01 of the Amazon RDS DataService SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by RDSData.

StatementTimeoutException

_StatementTimeoutException :: AsError a => Getting (First ServiceError) a ServiceError

The execution of the SQL statement timed out.

ForbiddenException

_ForbiddenException :: AsError a => Getting (First ServiceError) a ServiceError

There are insufficient privileges to make the call.

NotFoundException

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The resourceArn, secretArn, or transactionId value can't be found.

ServiceUnavailableError

_ServiceUnavailableError :: AsError a => Getting (First ServiceError) a ServiceError

The service specified by the resourceArn parameter is not available.

InternalServerErrorException

_InternalServerErrorException :: AsError a => Getting (First ServiceError) a ServiceError

An internal error occurred.

BadRequestException

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

There is an error in the call or in a SQL statement.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

RollbackTransaction

data RollbackTransaction

The request parameters represent the input of a request to perform a rollback of a transaction.

See: newRollbackTransaction smart constructor.

Constructors

RollbackTransaction' Text Text Text

newRollbackTransaction

Arguments

:: Text
$sel:resourceArn:RollbackTransaction'

-> Text
$sel:secretArn:RollbackTransaction'

-> Text
$sel:transactionId:RollbackTransaction'

-> RollbackTransaction

Create a value of RollbackTransaction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:RollbackTransaction', rollbackTransaction_resourceArn - The Amazon Resource Name (ARN) of the Aurora Serverless DB cluster.

$sel:secretArn:RollbackTransaction', rollbackTransaction_secretArn - The name or ARN of the secret that enables access to the DB cluster.

$sel:transactionId:RollbackTransaction', rollbackTransaction_transactionId - The identifier of the transaction to roll back.

data RollbackTransactionResponse

The response elements represent the output of a request to perform a rollback of a transaction.

See: newRollbackTransactionResponse smart constructor.

Constructors

RollbackTransactionResponse' (Maybe Text) Int

newRollbackTransactionResponse

Arguments

:: Int
$sel:httpStatus:RollbackTransactionResponse'

-> RollbackTransactionResponse

Create a value of RollbackTransactionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transactionStatus:RollbackTransactionResponse', rollbackTransactionResponse_transactionStatus - The status of the rollback operation.

$sel:httpStatus:RollbackTransactionResponse', rollbackTransactionResponse_httpStatus - The response's http status code.

BeginTransaction

data BeginTransaction

The request parameters represent the input of a request to start a SQL transaction.

See: newBeginTransaction smart constructor.

Constructors

BeginTransaction' (Maybe Text) (Maybe Text) Text Text

newBeginTransaction

Arguments

:: Text
$sel:resourceArn:BeginTransaction'

-> Text
$sel:secretArn:BeginTransaction'

-> BeginTransaction

Create a value of BeginTransaction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:database:BeginTransaction', beginTransaction_database - The name of the database.

$sel:schema:BeginTransaction', beginTransaction_schema - The name of the database schema.

$sel:resourceArn:BeginTransaction', beginTransaction_resourceArn - The Amazon Resource Name (ARN) of the Aurora Serverless DB cluster.

$sel:secretArn:BeginTransaction', beginTransaction_secretArn - The name or ARN of the secret that enables access to the DB cluster.

data BeginTransactionResponse

The response elements represent the output of a request to start a SQL transaction.

See: newBeginTransactionResponse smart constructor.

Constructors

BeginTransactionResponse' (Maybe Text) Int

newBeginTransactionResponse

Arguments

:: Int
$sel:httpStatus:BeginTransactionResponse'

-> BeginTransactionResponse

Create a value of BeginTransactionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transactionId:BeginTransactionResponse', beginTransactionResponse_transactionId - The transaction ID of the transaction started by the call.

$sel:httpStatus:BeginTransactionResponse', beginTransactionResponse_httpStatus - The response's http status code.

BatchExecuteStatement

data BatchExecuteStatement

The request parameters represent the input of a SQL statement over an array of data.

See: newBatchExecuteStatement smart constructor.

Constructors

BatchExecuteStatement' (Maybe Text) (Maybe [[SqlParameter]]) (Maybe Text) (Maybe Text) Text Text Text

newBatchExecuteStatement

Arguments

:: Text
$sel:resourceArn:BatchExecuteStatement'

-> Text
$sel:secretArn:BatchExecuteStatement'

-> Text
$sel:sql:BatchExecuteStatement'

-> BatchExecuteStatement

Create a value of BatchExecuteStatement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:database:BatchExecuteStatement', batchExecuteStatement_database - The name of the database.

$sel:parameterSets:BatchExecuteStatement', batchExecuteStatement_parameterSets - The parameter set for the batch operation.

The SQL statement is executed as many times as the number of parameter sets provided. To execute a SQL statement with no parameters, use one of the following options:

Specify one or more empty parameter sets.

Use the ExecuteStatement operation instead of the BatchExecuteStatement operation.

Array parameters are not supported.

$sel:transactionId:BatchExecuteStatement', batchExecuteStatement_transactionId - The identifier of a transaction that was started by using the BeginTransaction operation. Specify the transaction ID of the transaction that you want to include the SQL statement in.

If the SQL statement is not part of a transaction, don't set this parameter.

$sel:schema:BatchExecuteStatement', batchExecuteStatement_schema - The name of the database schema.

$sel:resourceArn:BatchExecuteStatement', batchExecuteStatement_resourceArn - The Amazon Resource Name (ARN) of the Aurora Serverless DB cluster.

$sel:secretArn:BatchExecuteStatement', batchExecuteStatement_secretArn - The name or ARN of the secret that enables access to the DB cluster.

$sel:sql:BatchExecuteStatement', batchExecuteStatement_sql - The SQL statement to run.

data BatchExecuteStatementResponse

The response elements represent the output of a SQL statement over an array of data.

See: newBatchExecuteStatementResponse smart constructor.

Constructors

BatchExecuteStatementResponse' (Maybe [UpdateResult]) Int

newBatchExecuteStatementResponse

Arguments

:: Int
$sel:httpStatus:BatchExecuteStatementResponse'

-> BatchExecuteStatementResponse

Create a value of BatchExecuteStatementResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateResults:BatchExecuteStatementResponse', batchExecuteStatementResponse_updateResults - The execution results of each batch entry.

$sel:httpStatus:BatchExecuteStatementResponse', batchExecuteStatementResponse_httpStatus - The response's http status code.

ExecuteStatement

data ExecuteStatement

The request parameters represent the input of a request to run a SQL statement against a database.

See: newExecuteStatement smart constructor.

Constructors

ExecuteStatement' (Maybe Text) (Maybe Text) (Maybe Text) (Maybe [SqlParameter]) (Maybe Bool) (Maybe ResultSetOptions) (Maybe Bool) Text Text Text

newExecuteStatement

Arguments

:: Text
$sel:resourceArn:ExecuteStatement'

-> Text
$sel:secretArn:ExecuteStatement'

-> Text
$sel:sql:ExecuteStatement'

-> ExecuteStatement

Create a value of ExecuteStatement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:database:ExecuteStatement', executeStatement_database - The name of the database.

$sel:transactionId:ExecuteStatement', executeStatement_transactionId - The identifier of a transaction that was started by using the BeginTransaction operation. Specify the transaction ID of the transaction that you want to include the SQL statement in.

If the SQL statement is not part of a transaction, don't set this parameter.

$sel:schema:ExecuteStatement', executeStatement_schema - The name of the database schema.

Currently, the schema parameter isn't supported.

$sel:parameters:ExecuteStatement', executeStatement_parameters - The parameters for the SQL statement.

Array parameters are not supported.

$sel:includeResultMetadata:ExecuteStatement', executeStatement_includeResultMetadata - A value that indicates whether to include metadata in the results.

$sel:resultSetOptions:ExecuteStatement', executeStatement_resultSetOptions - Options that control how the result set is returned.

$sel:continueAfterTimeout:ExecuteStatement', executeStatement_continueAfterTimeout - A value that indicates whether to continue running the statement after the call times out. By default, the statement stops running when the call times out.

For DDL statements, we recommend continuing to run the statement after the call times out. When a DDL statement terminates before it is finished running, it can result in errors and possibly corrupted data structures.

$sel:resourceArn:ExecuteStatement', executeStatement_resourceArn - The Amazon Resource Name (ARN) of the Aurora Serverless DB cluster.

$sel:secretArn:ExecuteStatement', executeStatement_secretArn - The name or ARN of the secret that enables access to the DB cluster.

$sel:sql:ExecuteStatement', executeStatement_sql - The SQL statement to run.

data ExecuteStatementResponse

The response elements represent the output of a request to run a SQL statement against a database.

See: newExecuteStatementResponse smart constructor.

Constructors

ExecuteStatementResponse' (Maybe [[Field]]) (Maybe [ColumnMetadata]) (Maybe [Field]) (Maybe Integer) Int

newExecuteStatementResponse

Arguments

:: Int
$sel:httpStatus:ExecuteStatementResponse'

-> ExecuteStatementResponse

Create a value of ExecuteStatementResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:records:ExecuteStatementResponse', executeStatementResponse_records - The records returned by the SQL statement.

$sel:columnMetadata:ExecuteStatementResponse', executeStatementResponse_columnMetadata - Metadata for the columns included in the results.

$sel:generatedFields:ExecuteStatementResponse', executeStatementResponse_generatedFields - Values for fields generated during the request.

 <note> <p>The <code>generatedFields</code> data isn't supported by Aurora PostgreSQL. To get the values of generated fields, use the <code>RETURNING</code> clause. For more information, see <a href="https://www.postgresql.org/docs/10/dml-returning.html">Returning Data From Modified Rows</a> in the PostgreSQL documentation.</p> </note>

$sel:numberOfRecordsUpdated:ExecuteStatementResponse', executeStatementResponse_numberOfRecordsUpdated - The number of records updated by the request.

$sel:httpStatus:ExecuteStatementResponse', executeStatementResponse_httpStatus - The response's http status code.

CommitTransaction

data CommitTransaction

The request parameters represent the input of a commit transaction request.

See: newCommitTransaction smart constructor.

Constructors

CommitTransaction' Text Text Text

newCommitTransaction

Arguments

:: Text
$sel:resourceArn:CommitTransaction'

-> Text
$sel:secretArn:CommitTransaction'

-> Text
$sel:transactionId:CommitTransaction'

-> CommitTransaction

Create a value of CommitTransaction with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:resourceArn:CommitTransaction', commitTransaction_resourceArn - The Amazon Resource Name (ARN) of the Aurora Serverless DB cluster.

$sel:secretArn:CommitTransaction', commitTransaction_secretArn - The name or ARN of the secret that enables access to the DB cluster.

$sel:transactionId:CommitTransaction', commitTransaction_transactionId - The identifier of the transaction to end and commit.

data CommitTransactionResponse

The response elements represent the output of a commit transaction request.

See: newCommitTransactionResponse smart constructor.

Constructors

CommitTransactionResponse' (Maybe Text) Int

newCommitTransactionResponse

Arguments

:: Int
$sel:httpStatus:CommitTransactionResponse'

-> CommitTransactionResponse

Create a value of CommitTransactionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:transactionStatus:CommitTransactionResponse', commitTransactionResponse_transactionStatus - The status of the commit operation.

$sel:httpStatus:CommitTransactionResponse', commitTransactionResponse_httpStatus - The response's http status code.

Types

DecimalReturnType

newtype DecimalReturnType

Constructors

DecimalReturnType'

Fields

fromDecimalReturnType :: Text

Bundled Patterns

pattern DecimalReturnType_DOUBLE_OR_LONG :: DecimalReturnType

pattern DecimalReturnType_STRING :: DecimalReturnType

TypeHint

newtype TypeHint

Constructors

TypeHint'

Fields

fromTypeHint :: Text

Bundled Patterns

pattern TypeHint_DATE :: TypeHint

pattern TypeHint_DECIMAL :: TypeHint

pattern TypeHint_JSON :: TypeHint

pattern TypeHint_TIME :: TypeHint

pattern TypeHint_TIMESTAMP :: TypeHint

pattern TypeHint_UUID :: TypeHint

ArrayValue

data ArrayValue

Contains an array.

See: newArrayValue smart constructor.

Constructors

ArrayValue' (Maybe [Integer]) (Maybe [Double]) (Maybe [Text]) (Maybe [ArrayValue]) (Maybe [Bool])

newArrayValue :: ArrayValue

Create a value of ArrayValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:longValues:ArrayValue', arrayValue_longValues - An array of floating point numbers.

$sel:doubleValues:ArrayValue', arrayValue_doubleValues - An array of integers.

$sel:stringValues:ArrayValue', arrayValue_stringValues - An array of strings.

$sel:arrayValues:ArrayValue', arrayValue_arrayValues - An array of arrays.

$sel:booleanValues:ArrayValue', arrayValue_booleanValues - An array of Boolean values.

ColumnMetadata

data ColumnMetadata

Contains the metadata for a column.

See: newColumnMetadata smart constructor.

Constructors

ColumnMetadata' (Maybe Text) (Maybe Bool) (Maybe Bool) (Maybe Int) (Maybe Int) (Maybe Text) (Maybe Bool) (Maybe Text) (Maybe Int) (Maybe Int) (Maybe Bool) (Maybe Text) (Maybe Int) (Maybe Text)

newColumnMetadata :: ColumnMetadata

Create a value of ColumnMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeName:ColumnMetadata', columnMetadata_typeName - The database-specific data type of the column.

$sel:isCaseSensitive:ColumnMetadata', columnMetadata_isCaseSensitive - A value that indicates whether the column is case-sensitive.

$sel:isCurrency:ColumnMetadata', columnMetadata_isCurrency - A value that indicates whether the column contains currency values.

$sel:scale:ColumnMetadata', columnMetadata_scale - The scale value of a decimal number column.

$sel:precision:ColumnMetadata', columnMetadata_precision - The precision value of a decimal number column.

$sel:schemaName:ColumnMetadata', columnMetadata_schemaName - The name of the schema that owns the table that includes the column.

$sel:isAutoIncrement:ColumnMetadata', columnMetadata_isAutoIncrement - A value that indicates whether the column increments automatically.

$sel:name:ColumnMetadata', columnMetadata_name - The name of the column.

$sel:arrayBaseColumnType:ColumnMetadata', columnMetadata_arrayBaseColumnType - The type of the column.

$sel:type':ColumnMetadata', columnMetadata_type - The type of the column.

$sel:isSigned:ColumnMetadata', columnMetadata_isSigned - A value that indicates whether an integer column is signed.

$sel:label:ColumnMetadata', columnMetadata_label - The label for the column.

$sel:nullable:ColumnMetadata', columnMetadata_nullable - A value that indicates whether the column is nullable.

$sel:tableName:ColumnMetadata', columnMetadata_tableName - The name of the table that includes the column.

Field

data Field

Contains a value.

See: newField smart constructor.

Constructors

Field' (Maybe Double) (Maybe Text) (Maybe Integer) (Maybe Bool) (Maybe ArrayValue) (Maybe Base64) (Maybe Bool)

newField :: Field

Create a value of Field with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:doubleValue:Field', field_doubleValue - A value of double data type.

$sel:stringValue:Field', field_stringValue - A value of string data type.

$sel:longValue:Field', field_longValue - A value of long data type.

$sel:booleanValue:Field', field_booleanValue - A value of Boolean data type.

$sel:arrayValue:Field', field_arrayValue - An array of values.

$sel:blobValue:Field', field_blobValue - A value of BLOB data type.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:isNull:Field', field_isNull - A NULL value.

ResultSetOptions

data ResultSetOptions

Options that control how the result set is returned.

See: newResultSetOptions smart constructor.

Constructors

ResultSetOptions' (Maybe DecimalReturnType)

newResultSetOptions :: ResultSetOptions

Create a value of ResultSetOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:decimalReturnType:ResultSetOptions', resultSetOptions_decimalReturnType - A value that indicates how a field of DECIMAL type is represented in the response. The value of STRING, the default, specifies that it is converted to a String value. The value of DOUBLE_OR_LONG specifies that it is converted to a Long value if its scale is 0, or to a Double value otherwise.

Conversion to Double or Long can result in roundoff errors due to precision loss. We recommend converting to String, especially when working with currency values.

SqlParameter

data SqlParameter

A parameter used in a SQL statement.

See: newSqlParameter smart constructor.

Constructors

SqlParameter' (Maybe Field) (Maybe Text) (Maybe TypeHint)

newSqlParameter :: SqlParameter

Create a value of SqlParameter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:SqlParameter', sqlParameter_value - The value of the parameter.

$sel:name:SqlParameter', sqlParameter_name - The name of the parameter.

$sel:typeHint:SqlParameter', sqlParameter_typeHint - A hint that specifies the correct object type for data type mapping. Possible values are as follows:

DATE - The corresponding String parameter value is sent as an object of DATE type to the database. The accepted format is YYYY-MM-DD.

DECIMAL - The corresponding String parameter value is sent as an object of DECIMAL type to the database.

JSON - The corresponding String parameter value is sent as an object of JSON type to the database.

TIME - The corresponding String parameter value is sent as an object of TIME type to the database. The accepted format is HH:MM:SS[.FFF].

TIMESTAMP - The corresponding String parameter value is sent as an object of TIMESTAMP type to the database. The accepted format is YYYY-MM-DD HH:MM:SS[.FFF].

UUID - The corresponding String parameter value is sent as an object of UUID type to the database.

UpdateResult

data UpdateResult

The response elements represent the results of an update.

See: newUpdateResult smart constructor.

Constructors

UpdateResult' (Maybe [Field])

newUpdateResult :: UpdateResult

Create a value of UpdateResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:generatedFields:UpdateResult', updateResult_generatedFields - Values for fields generated during the request.