Network.AWS.RDSData.BatchExecuteStatement

Description

Runs a batch SQL statement over an array of data.

You can run bulk update and insert operations for multiple records using a DML statement with different parameter sets. Bulk operations can provide a significant performance improvement over individual insert and update operations.

If a call isn't part of a transaction because it doesn't include the transactionID parameter, changes that result from the call are committed automatically.

Creating a Request

data BatchExecuteStatement

The request parameters represent the input of a SQL statement over an array of data.

See: newBatchExecuteStatement smart constructor.

Constructors

BatchExecuteStatement'

Fields

database :: Maybe Text

The name of the database.

parameterSets :: Maybe [[SqlParameter]]

The parameter set for the batch operation.

The SQL statement is executed as many times as the number of parameter sets provided. To execute a SQL statement with no parameters, use one of the following options:

Specify one or more empty parameter sets.

Use the ExecuteStatement operation instead of the BatchExecuteStatement operation.

Array parameters are not supported.

transactionId :: Maybe Text

The identifier of a transaction that was started by using the BeginTransaction operation. Specify the transaction ID of the transaction that you want to include the SQL statement in.

If the SQL statement is not part of a transaction, don't set this parameter.

schema :: Maybe Text

The name of the database schema.

resourceArn :: Text

The Amazon Resource Name (ARN) of the Aurora Serverless DB cluster.

secretArn :: Text

The name or ARN of the secret that enables access to the DB cluster.

sql :: Text

The SQL statement to run.

newBatchExecuteStatement

Arguments

:: Text
$sel:resourceArn:BatchExecuteStatement'

-> Text
$sel:secretArn:BatchExecuteStatement'

-> Text
$sel:sql:BatchExecuteStatement'

-> BatchExecuteStatement

Create a value of BatchExecuteStatement with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:database:BatchExecuteStatement', batchExecuteStatement_database - The name of the database.

$sel:parameterSets:BatchExecuteStatement', batchExecuteStatement_parameterSets - The parameter set for the batch operation.

The SQL statement is executed as many times as the number of parameter sets provided. To execute a SQL statement with no parameters, use one of the following options:

Specify one or more empty parameter sets.

Use the ExecuteStatement operation instead of the BatchExecuteStatement operation.

Array parameters are not supported.

$sel:transactionId:BatchExecuteStatement', batchExecuteStatement_transactionId - The identifier of a transaction that was started by using the BeginTransaction operation. Specify the transaction ID of the transaction that you want to include the SQL statement in.

If the SQL statement is not part of a transaction, don't set this parameter.

$sel:schema:BatchExecuteStatement', batchExecuteStatement_schema - The name of the database schema.

$sel:resourceArn:BatchExecuteStatement', batchExecuteStatement_resourceArn - The Amazon Resource Name (ARN) of the Aurora Serverless DB cluster.

$sel:secretArn:BatchExecuteStatement', batchExecuteStatement_secretArn - The name or ARN of the secret that enables access to the DB cluster.

$sel:sql:BatchExecuteStatement', batchExecuteStatement_sql - The SQL statement to run.

Request Lenses

batchExecuteStatement_database :: Lens' BatchExecuteStatement (Maybe Text)

The name of the database.

batchExecuteStatement_parameterSets :: Lens' BatchExecuteStatement (Maybe [[SqlParameter]])

The parameter set for the batch operation.

The SQL statement is executed as many times as the number of parameter sets provided. To execute a SQL statement with no parameters, use one of the following options:

Specify one or more empty parameter sets.

Use the ExecuteStatement operation instead of the BatchExecuteStatement operation.

Array parameters are not supported.

batchExecuteStatement_transactionId :: Lens' BatchExecuteStatement (Maybe Text)

The identifier of a transaction that was started by using the BeginTransaction operation. Specify the transaction ID of the transaction that you want to include the SQL statement in.

If the SQL statement is not part of a transaction, don't set this parameter.

batchExecuteStatement_schema :: Lens' BatchExecuteStatement (Maybe Text)

The name of the database schema.

batchExecuteStatement_resourceArn :: Lens' BatchExecuteStatement Text

The Amazon Resource Name (ARN) of the Aurora Serverless DB cluster.

batchExecuteStatement_secretArn :: Lens' BatchExecuteStatement Text

The name or ARN of the secret that enables access to the DB cluster.

batchExecuteStatement_sql :: Lens' BatchExecuteStatement Text

The SQL statement to run.

Destructuring the Response

data BatchExecuteStatementResponse

The response elements represent the output of a SQL statement over an array of data.

See: newBatchExecuteStatementResponse smart constructor.

Constructors

BatchExecuteStatementResponse'

Fields

updateResults :: Maybe [UpdateResult]

The execution results of each batch entry.

httpStatus :: Int

The response's http status code.

newBatchExecuteStatementResponse

Arguments

:: Int
$sel:httpStatus:BatchExecuteStatementResponse'

-> BatchExecuteStatementResponse

Create a value of BatchExecuteStatementResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:updateResults:BatchExecuteStatementResponse', batchExecuteStatementResponse_updateResults - The execution results of each batch entry.

$sel:httpStatus:BatchExecuteStatementResponse', batchExecuteStatementResponse_httpStatus - The response's http status code.

Response Lenses

batchExecuteStatementResponse_updateResults :: Lens' BatchExecuteStatementResponse (Maybe [UpdateResult])

The execution results of each batch entry.

batchExecuteStatementResponse_httpStatus :: Lens' BatchExecuteStatementResponse Int

The response's http status code.