Network.AWS.RDSData.Types

Description

Service Configuration

defaultService :: Service

API version 2018-08-01 of the Amazon RDS DataService SDK configuration.

Errors

_StatementTimeoutException :: AsError a => Getting (First ServiceError) a ServiceError

The execution of the SQL statement timed out.

_ForbiddenException :: AsError a => Getting (First ServiceError) a ServiceError

There are insufficient privileges to make the call.

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

The resourceArn, secretArn, or transactionId value can't be found.

_ServiceUnavailableError :: AsError a => Getting (First ServiceError) a ServiceError

The service specified by the resourceArn parameter is not available.

_InternalServerErrorException :: AsError a => Getting (First ServiceError) a ServiceError

An internal error occurred.

_BadRequestException :: AsError a => Getting (First ServiceError) a ServiceError

There is an error in the call or in a SQL statement.

DecimalReturnType

newtype DecimalReturnType

Constructors

DecimalReturnType'

Fields

fromDecimalReturnType :: Text

Bundled Patterns

pattern DecimalReturnType_DOUBLE_OR_LONG :: DecimalReturnType

pattern DecimalReturnType_STRING :: DecimalReturnType

TypeHint

newtype TypeHint

Constructors

TypeHint'

Fields

fromTypeHint :: Text

Bundled Patterns

pattern TypeHint_DATE :: TypeHint

pattern TypeHint_DECIMAL :: TypeHint

pattern TypeHint_JSON :: TypeHint

pattern TypeHint_TIME :: TypeHint

pattern TypeHint_TIMESTAMP :: TypeHint

pattern TypeHint_UUID :: TypeHint

ArrayValue

data ArrayValue

Contains an array.

See: newArrayValue smart constructor.

Constructors

ArrayValue'

Fields

longValues :: Maybe [Integer]

An array of floating point numbers.

doubleValues :: Maybe [Double]

An array of integers.

stringValues :: Maybe [Text]

An array of strings.

arrayValues :: Maybe [ArrayValue]

An array of arrays.

booleanValues :: Maybe [Bool]

An array of Boolean values.

newArrayValue :: ArrayValue

Create a value of ArrayValue with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:longValues:ArrayValue', arrayValue_longValues - An array of floating point numbers.

$sel:doubleValues:ArrayValue', arrayValue_doubleValues - An array of integers.

$sel:stringValues:ArrayValue', arrayValue_stringValues - An array of strings.

$sel:arrayValues:ArrayValue', arrayValue_arrayValues - An array of arrays.

$sel:booleanValues:ArrayValue', arrayValue_booleanValues - An array of Boolean values.

arrayValue_longValues :: Lens' ArrayValue (Maybe [Integer])

An array of floating point numbers.

arrayValue_doubleValues :: Lens' ArrayValue (Maybe [Double])

An array of integers.

arrayValue_stringValues :: Lens' ArrayValue (Maybe [Text])

An array of strings.

arrayValue_arrayValues :: Lens' ArrayValue (Maybe [ArrayValue])

An array of arrays.

arrayValue_booleanValues :: Lens' ArrayValue (Maybe [Bool])

An array of Boolean values.

ColumnMetadata

data ColumnMetadata

Contains the metadata for a column.

See: newColumnMetadata smart constructor.

Constructors

ColumnMetadata'

Fields

typeName :: Maybe Text

The database-specific data type of the column.

isCaseSensitive :: Maybe Bool

A value that indicates whether the column is case-sensitive.

isCurrency :: Maybe Bool

A value that indicates whether the column contains currency values.

scale :: Maybe Int

The scale value of a decimal number column.

precision :: Maybe Int

The precision value of a decimal number column.

schemaName :: Maybe Text

The name of the schema that owns the table that includes the column.

isAutoIncrement :: Maybe Bool

A value that indicates whether the column increments automatically.

name :: Maybe Text

The name of the column.

arrayBaseColumnType :: Maybe Int

The type of the column.

type' :: Maybe Int

The type of the column.

isSigned :: Maybe Bool

A value that indicates whether an integer column is signed.

label :: Maybe Text

The label for the column.

nullable :: Maybe Int

A value that indicates whether the column is nullable.

tableName :: Maybe Text

The name of the table that includes the column.

newColumnMetadata :: ColumnMetadata

Create a value of ColumnMetadata with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:typeName:ColumnMetadata', columnMetadata_typeName - The database-specific data type of the column.

$sel:isCaseSensitive:ColumnMetadata', columnMetadata_isCaseSensitive - A value that indicates whether the column is case-sensitive.

$sel:isCurrency:ColumnMetadata', columnMetadata_isCurrency - A value that indicates whether the column contains currency values.

$sel:scale:ColumnMetadata', columnMetadata_scale - The scale value of a decimal number column.

$sel:precision:ColumnMetadata', columnMetadata_precision - The precision value of a decimal number column.

$sel:schemaName:ColumnMetadata', columnMetadata_schemaName - The name of the schema that owns the table that includes the column.

$sel:isAutoIncrement:ColumnMetadata', columnMetadata_isAutoIncrement - A value that indicates whether the column increments automatically.

$sel:name:ColumnMetadata', columnMetadata_name - The name of the column.

$sel:arrayBaseColumnType:ColumnMetadata', columnMetadata_arrayBaseColumnType - The type of the column.

$sel:type':ColumnMetadata', columnMetadata_type - The type of the column.

$sel:isSigned:ColumnMetadata', columnMetadata_isSigned - A value that indicates whether an integer column is signed.

$sel:label:ColumnMetadata', columnMetadata_label - The label for the column.

$sel:nullable:ColumnMetadata', columnMetadata_nullable - A value that indicates whether the column is nullable.

$sel:tableName:ColumnMetadata', columnMetadata_tableName - The name of the table that includes the column.

columnMetadata_typeName :: Lens' ColumnMetadata (Maybe Text)

The database-specific data type of the column.

columnMetadata_isCaseSensitive :: Lens' ColumnMetadata (Maybe Bool)

A value that indicates whether the column is case-sensitive.

columnMetadata_isCurrency :: Lens' ColumnMetadata (Maybe Bool)

A value that indicates whether the column contains currency values.

columnMetadata_scale :: Lens' ColumnMetadata (Maybe Int)

The scale value of a decimal number column.

columnMetadata_precision :: Lens' ColumnMetadata (Maybe Int)

The precision value of a decimal number column.

columnMetadata_schemaName :: Lens' ColumnMetadata (Maybe Text)

The name of the schema that owns the table that includes the column.

columnMetadata_isAutoIncrement :: Lens' ColumnMetadata (Maybe Bool)

A value that indicates whether the column increments automatically.

columnMetadata_name :: Lens' ColumnMetadata (Maybe Text)

The name of the column.

columnMetadata_arrayBaseColumnType :: Lens' ColumnMetadata (Maybe Int)

The type of the column.

columnMetadata_type :: Lens' ColumnMetadata (Maybe Int)

The type of the column.

columnMetadata_isSigned :: Lens' ColumnMetadata (Maybe Bool)

A value that indicates whether an integer column is signed.

columnMetadata_label :: Lens' ColumnMetadata (Maybe Text)

The label for the column.

columnMetadata_nullable :: Lens' ColumnMetadata (Maybe Int)

A value that indicates whether the column is nullable.

columnMetadata_tableName :: Lens' ColumnMetadata (Maybe Text)

The name of the table that includes the column.

Field

data Field

Contains a value.

See: newField smart constructor.

Constructors

Field'

Fields

doubleValue :: Maybe Double

A value of double data type.

stringValue :: Maybe Text

A value of string data type.

longValue :: Maybe Integer

A value of long data type.

booleanValue :: Maybe Bool

A value of Boolean data type.

arrayValue :: Maybe ArrayValue

An array of values.

blobValue :: Maybe Base64

A value of BLOB data type.

isNull :: Maybe Bool

A NULL value.

newField :: Field

Create a value of Field with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:doubleValue:Field', field_doubleValue - A value of double data type.

$sel:stringValue:Field', field_stringValue - A value of string data type.

$sel:longValue:Field', field_longValue - A value of long data type.

$sel:booleanValue:Field', field_booleanValue - A value of Boolean data type.

$sel:arrayValue:Field', field_arrayValue - An array of values.

$sel:blobValue:Field', field_blobValue - A value of BLOB data type.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

$sel:isNull:Field', field_isNull - A NULL value.

field_doubleValue :: Lens' Field (Maybe Double)

A value of double data type.

field_stringValue :: Lens' Field (Maybe Text)

A value of string data type.

field_longValue :: Lens' Field (Maybe Integer)

A value of long data type.

field_booleanValue :: Lens' Field (Maybe Bool)

A value of Boolean data type.

field_arrayValue :: Lens' Field (Maybe ArrayValue)

An array of values.

field_blobValue :: Lens' Field (Maybe ByteString)

A value of BLOB data type.-- -- Note: This Lens automatically encodes and decodes Base64 data. -- The underlying isomorphism will encode to Base64 representation during -- serialisation, and decode from Base64 representation during deserialisation. -- This Lens accepts and returns only raw unencoded data.

field_isNull :: Lens' Field (Maybe Bool)

A NULL value.

ResultSetOptions

data ResultSetOptions

Options that control how the result set is returned.

See: newResultSetOptions smart constructor.

Constructors

ResultSetOptions'

Fields

decimalReturnType :: Maybe DecimalReturnType

A value that indicates how a field of DECIMAL type is represented in the response. The value of STRING, the default, specifies that it is converted to a String value. The value of DOUBLE_OR_LONG specifies that it is converted to a Long value if its scale is 0, or to a Double value otherwise.

Conversion to Double or Long can result in roundoff errors due to precision loss. We recommend converting to String, especially when working with currency values.

newResultSetOptions :: ResultSetOptions

Create a value of ResultSetOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:decimalReturnType:ResultSetOptions', resultSetOptions_decimalReturnType - A value that indicates how a field of DECIMAL type is represented in the response. The value of STRING, the default, specifies that it is converted to a String value. The value of DOUBLE_OR_LONG specifies that it is converted to a Long value if its scale is 0, or to a Double value otherwise.

Conversion to Double or Long can result in roundoff errors due to precision loss. We recommend converting to String, especially when working with currency values.

resultSetOptions_decimalReturnType :: Lens' ResultSetOptions (Maybe DecimalReturnType)

A value that indicates how a field of DECIMAL type is represented in the response. The value of STRING, the default, specifies that it is converted to a String value. The value of DOUBLE_OR_LONG specifies that it is converted to a Long value if its scale is 0, or to a Double value otherwise.

Conversion to Double or Long can result in roundoff errors due to precision loss. We recommend converting to String, especially when working with currency values.

SqlParameter

data SqlParameter

A parameter used in a SQL statement.

See: newSqlParameter smart constructor.

Constructors

SqlParameter'

Fields

value :: Maybe Field

The value of the parameter.

name :: Maybe Text

The name of the parameter.

typeHint :: Maybe TypeHint

A hint that specifies the correct object type for data type mapping. Possible values are as follows:

DATE - The corresponding String parameter value is sent as an object of DATE type to the database. The accepted format is YYYY-MM-DD.

DECIMAL - The corresponding String parameter value is sent as an object of DECIMAL type to the database.

JSON - The corresponding String parameter value is sent as an object of JSON type to the database.

TIME - The corresponding String parameter value is sent as an object of TIME type to the database. The accepted format is HH:MM:SS[.FFF].

TIMESTAMP - The corresponding String parameter value is sent as an object of TIMESTAMP type to the database. The accepted format is YYYY-MM-DD HH:MM:SS[.FFF].

UUID - The corresponding String parameter value is sent as an object of UUID type to the database.

newSqlParameter :: SqlParameter

Create a value of SqlParameter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:SqlParameter', sqlParameter_value - The value of the parameter.

$sel:name:SqlParameter', sqlParameter_name - The name of the parameter.

$sel:typeHint:SqlParameter', sqlParameter_typeHint - A hint that specifies the correct object type for data type mapping. Possible values are as follows:

DATE - The corresponding String parameter value is sent as an object of DATE type to the database. The accepted format is YYYY-MM-DD.

DECIMAL - The corresponding String parameter value is sent as an object of DECIMAL type to the database.

JSON - The corresponding String parameter value is sent as an object of JSON type to the database.

TIME - The corresponding String parameter value is sent as an object of TIME type to the database. The accepted format is HH:MM:SS[.FFF].

TIMESTAMP - The corresponding String parameter value is sent as an object of TIMESTAMP type to the database. The accepted format is YYYY-MM-DD HH:MM:SS[.FFF].

UUID - The corresponding String parameter value is sent as an object of UUID type to the database.

sqlParameter_value :: Lens' SqlParameter (Maybe Field)

The value of the parameter.

sqlParameter_name :: Lens' SqlParameter (Maybe Text)

The name of the parameter.

sqlParameter_typeHint :: Lens' SqlParameter (Maybe TypeHint)

A hint that specifies the correct object type for data type mapping. Possible values are as follows:

DATE - The corresponding String parameter value is sent as an object of DATE type to the database. The accepted format is YYYY-MM-DD.

DECIMAL - The corresponding String parameter value is sent as an object of DECIMAL type to the database.

JSON - The corresponding String parameter value is sent as an object of JSON type to the database.

TIME - The corresponding String parameter value is sent as an object of TIME type to the database. The accepted format is HH:MM:SS[.FFF].

TIMESTAMP - The corresponding String parameter value is sent as an object of TIMESTAMP type to the database. The accepted format is YYYY-MM-DD HH:MM:SS[.FFF].

UUID - The corresponding String parameter value is sent as an object of UUID type to the database.

UpdateResult

data UpdateResult

The response elements represent the results of an update.

See: newUpdateResult smart constructor.

Constructors

UpdateResult'

Fields

generatedFields :: Maybe [Field]

Values for fields generated during the request.

newUpdateResult :: UpdateResult

Create a value of UpdateResult with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:generatedFields:UpdateResult', updateResult_generatedFields - Values for fields generated during the request.

updateResult_generatedFields :: Lens' UpdateResult (Maybe [Field])

Values for fields generated during the request.