Network.AWS.RDSData.Types.SqlParameter

Description

Documentation

data SqlParameter

A parameter used in a SQL statement.

See: newSqlParameter smart constructor.

Constructors

SqlParameter'

Fields

value :: Maybe Field

The value of the parameter.

name :: Maybe Text

The name of the parameter.

typeHint :: Maybe TypeHint

A hint that specifies the correct object type for data type mapping. Possible values are as follows:

DATE - The corresponding String parameter value is sent as an object of DATE type to the database. The accepted format is YYYY-MM-DD.

DECIMAL - The corresponding String parameter value is sent as an object of DECIMAL type to the database.

JSON - The corresponding String parameter value is sent as an object of JSON type to the database.

TIME - The corresponding String parameter value is sent as an object of TIME type to the database. The accepted format is HH:MM:SS[.FFF].

TIMESTAMP - The corresponding String parameter value is sent as an object of TIMESTAMP type to the database. The accepted format is YYYY-MM-DD HH:MM:SS[.FFF].

UUID - The corresponding String parameter value is sent as an object of UUID type to the database.

newSqlParameter :: SqlParameter

Create a value of SqlParameter with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:value:SqlParameter', sqlParameter_value - The value of the parameter.

$sel:name:SqlParameter', sqlParameter_name - The name of the parameter.

$sel:typeHint:SqlParameter', sqlParameter_typeHint - A hint that specifies the correct object type for data type mapping. Possible values are as follows:

DATE - The corresponding String parameter value is sent as an object of DATE type to the database. The accepted format is YYYY-MM-DD.

DECIMAL - The corresponding String parameter value is sent as an object of DECIMAL type to the database.

JSON - The corresponding String parameter value is sent as an object of JSON type to the database.

TIME - The corresponding String parameter value is sent as an object of TIME type to the database. The accepted format is HH:MM:SS[.FFF].

TIMESTAMP - The corresponding String parameter value is sent as an object of TIMESTAMP type to the database. The accepted format is YYYY-MM-DD HH:MM:SS[.FFF].

UUID - The corresponding String parameter value is sent as an object of UUID type to the database.

sqlParameter_value :: Lens' SqlParameter (Maybe Field)

The value of the parameter.

sqlParameter_name :: Lens' SqlParameter (Maybe Text)

The name of the parameter.

sqlParameter_typeHint :: Lens' SqlParameter (Maybe TypeHint)

A hint that specifies the correct object type for data type mapping. Possible values are as follows:

DATE - The corresponding String parameter value is sent as an object of DATE type to the database. The accepted format is YYYY-MM-DD.

DECIMAL - The corresponding String parameter value is sent as an object of DECIMAL type to the database.

JSON - The corresponding String parameter value is sent as an object of JSON type to the database.

TIME - The corresponding String parameter value is sent as an object of TIME type to the database. The accepted format is HH:MM:SS[.FFF].

TIMESTAMP - The corresponding String parameter value is sent as an object of TIMESTAMP type to the database. The accepted format is YYYY-MM-DD HH:MM:SS[.FFF].

UUID - The corresponding String parameter value is sent as an object of UUID type to the database.