Network.AWS.Redshift.AssociateDataShareConsumer

Description

From a datashare consumer account, associates a datashare with the account (AssociateEntireAccount) or the specified namespace (ConsumerArn). If you make this association, the consumer can consume the datashare.

Creating a Request

data AssociateDataShareConsumer

See: newAssociateDataShareConsumer smart constructor.

Constructors

AssociateDataShareConsumer'

Fields

associateEntireAccount :: Maybe Bool

A value that specifies whether the datashare is associated with the entire account.

consumerArn :: Maybe Text

The Amazon Resource Name (ARN) of the consumer that is associated with the datashare.

dataShareArn :: Text

The Amazon Resource Name (ARN) of the datashare that the consumer is to use with the account or the namespace.

newAssociateDataShareConsumer

Arguments

:: Text
$sel:dataShareArn:AssociateDataShareConsumer'

-> AssociateDataShareConsumer

Create a value of AssociateDataShareConsumer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:associateEntireAccount:AssociateDataShareConsumer', associateDataShareConsumer_associateEntireAccount - A value that specifies whether the datashare is associated with the entire account.

$sel:consumerArn:AssociateDataShareConsumer', associateDataShareConsumer_consumerArn - The Amazon Resource Name (ARN) of the consumer that is associated with the datashare.

$sel:dataShareArn:AssociateDataShareConsumer', associateDataShareConsumer_dataShareArn - The Amazon Resource Name (ARN) of the datashare that the consumer is to use with the account or the namespace.

Request Lenses

associateDataShareConsumer_associateEntireAccount :: Lens' AssociateDataShareConsumer (Maybe Bool)

A value that specifies whether the datashare is associated with the entire account.

associateDataShareConsumer_consumerArn :: Lens' AssociateDataShareConsumer (Maybe Text)

The Amazon Resource Name (ARN) of the consumer that is associated with the datashare.

associateDataShareConsumer_dataShareArn :: Lens' AssociateDataShareConsumer Text

The Amazon Resource Name (ARN) of the datashare that the consumer is to use with the account or the namespace.

Destructuring the Response

data DataShare

See: newDataShare smart constructor.

Constructors

DataShare'

Fields

dataShareAssociations :: Maybe [DataShareAssociation]

A value that specifies when the datashare has an association between a producer and data consumers.

allowPubliclyAccessibleConsumers :: Maybe Bool

A value that specifies whether the datashare can be shared to a publicly accessible cluster.

producerArn :: Maybe Text

The Amazon Resource Name (ARN) of the producer.

dataShareArn :: Maybe Text

An Amazon Resource Name (ARN) that references the datashare that is owned by a specific namespace of the producer cluster. A datashare ARN is in the arn:aws:redshift:{region}:{account-id}:{datashare}:{namespace-guid}/{datashare-name} format.

newDataShare :: DataShare

Create a value of DataShare with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataShareAssociations:DataShare', dataShare_dataShareAssociations - A value that specifies when the datashare has an association between a producer and data consumers.

$sel:allowPubliclyAccessibleConsumers:DataShare', dataShare_allowPubliclyAccessibleConsumers - A value that specifies whether the datashare can be shared to a publicly accessible cluster.

$sel:producerArn:DataShare', dataShare_producerArn - The Amazon Resource Name (ARN) of the producer.

$sel:dataShareArn:DataShare', dataShare_dataShareArn - An Amazon Resource Name (ARN) that references the datashare that is owned by a specific namespace of the producer cluster. A datashare ARN is in the arn:aws:redshift:{region}:{account-id}:{datashare}:{namespace-guid}/{datashare-name} format.

Response Lenses

dataShare_dataShareAssociations :: Lens' DataShare (Maybe [DataShareAssociation])

A value that specifies when the datashare has an association between a producer and data consumers.

dataShare_allowPubliclyAccessibleConsumers :: Lens' DataShare (Maybe Bool)

A value that specifies whether the datashare can be shared to a publicly accessible cluster.

dataShare_producerArn :: Lens' DataShare (Maybe Text)

The Amazon Resource Name (ARN) of the producer.

dataShare_dataShareArn :: Lens' DataShare (Maybe Text)

An Amazon Resource Name (ARN) that references the datashare that is owned by a specific namespace of the producer cluster. A datashare ARN is in the arn:aws:redshift:{region}:{account-id}:{datashare}:{namespace-guid}/{datashare-name} format.