Network.AWS.Redshift.AuthorizeClusterSecurityGroupIngress

Description

Adds an inbound (ingress) rule to an Amazon Redshift security group. Depending on whether the application accessing your cluster is running on the Internet or an Amazon EC2 instance, you can authorize inbound access to either a Classless Interdomain Routing (CIDR)/Internet Protocol (IP) range or to an Amazon EC2 security group. You can add as many as 20 ingress rules to an Amazon Redshift security group.

If you authorize access to an Amazon EC2 security group, specify EC2SecurityGroupName and EC2SecurityGroupOwnerId. The Amazon EC2 security group and Amazon Redshift cluster must be in the same Amazon Web Services Region.

If you authorize access to a CIDR/IP address range, specify CIDRIP. For an overview of CIDR blocks, see the Wikipedia article on Classless Inter-Domain Routing.

You must also associate the security group with a cluster so that clients running on these IP addresses or the EC2 instance are authorized to connect to the cluster. For information about managing security groups, go to Working with Security Groups in the Amazon Redshift Cluster Management Guide.

Creating a Request

data AuthorizeClusterSecurityGroupIngress

See: newAuthorizeClusterSecurityGroupIngress smart constructor.

Constructors

AuthorizeClusterSecurityGroupIngress'

Fields

eC2SecurityGroupOwnerId :: Maybe Text

The Amazon Web Services account number of the owner of the security group specified by the EC2SecurityGroupName parameter. The Amazon Web Services Access Key ID is not an acceptable value.

Example: 111122223333

eC2SecurityGroupName :: Maybe Text

The EC2 security group to be added the Amazon Redshift security group.

cidrip :: Maybe Text

The IP range to be added the Amazon Redshift security group.

clusterSecurityGroupName :: Text

The name of the security group to which the ingress rule is added.

newAuthorizeClusterSecurityGroupIngress

Arguments

:: Text
$sel:clusterSecurityGroupName:AuthorizeClusterSecurityGroupIngress'

-> AuthorizeClusterSecurityGroupIngress

Create a value of AuthorizeClusterSecurityGroupIngress with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eC2SecurityGroupOwnerId:AuthorizeClusterSecurityGroupIngress', authorizeClusterSecurityGroupIngress_eC2SecurityGroupOwnerId - The Amazon Web Services account number of the owner of the security group specified by the EC2SecurityGroupName parameter. The Amazon Web Services Access Key ID is not an acceptable value.

Example: 111122223333

$sel:eC2SecurityGroupName:AuthorizeClusterSecurityGroupIngress', authorizeClusterSecurityGroupIngress_eC2SecurityGroupName - The EC2 security group to be added the Amazon Redshift security group.

$sel:cidrip:AuthorizeClusterSecurityGroupIngress', authorizeClusterSecurityGroupIngress_cidrip - The IP range to be added the Amazon Redshift security group.

$sel:clusterSecurityGroupName:AuthorizeClusterSecurityGroupIngress', authorizeClusterSecurityGroupIngress_clusterSecurityGroupName - The name of the security group to which the ingress rule is added.

Request Lenses

authorizeClusterSecurityGroupIngress_eC2SecurityGroupOwnerId :: Lens' AuthorizeClusterSecurityGroupIngress (Maybe Text)

The Amazon Web Services account number of the owner of the security group specified by the EC2SecurityGroupName parameter. The Amazon Web Services Access Key ID is not an acceptable value.

Example: 111122223333

authorizeClusterSecurityGroupIngress_eC2SecurityGroupName :: Lens' AuthorizeClusterSecurityGroupIngress (Maybe Text)

The EC2 security group to be added the Amazon Redshift security group.

authorizeClusterSecurityGroupIngress_cidrip :: Lens' AuthorizeClusterSecurityGroupIngress (Maybe Text)

The IP range to be added the Amazon Redshift security group.

authorizeClusterSecurityGroupIngress_clusterSecurityGroupName :: Lens' AuthorizeClusterSecurityGroupIngress Text

The name of the security group to which the ingress rule is added.

Destructuring the Response

data AuthorizeClusterSecurityGroupIngressResponse

See: newAuthorizeClusterSecurityGroupIngressResponse smart constructor.

Constructors

AuthorizeClusterSecurityGroupIngressResponse'

Fields

clusterSecurityGroup :: Maybe ClusterSecurityGroup

httpStatus :: Int

The response's http status code.

newAuthorizeClusterSecurityGroupIngressResponse

Arguments

:: Int
$sel:httpStatus:AuthorizeClusterSecurityGroupIngressResponse'

-> AuthorizeClusterSecurityGroupIngressResponse

Create a value of AuthorizeClusterSecurityGroupIngressResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterSecurityGroup:AuthorizeClusterSecurityGroupIngressResponse', authorizeClusterSecurityGroupIngressResponse_clusterSecurityGroup - Undocumented member.

$sel:httpStatus:AuthorizeClusterSecurityGroupIngressResponse', authorizeClusterSecurityGroupIngressResponse_httpStatus - The response's http status code.

Response Lenses

authorizeClusterSecurityGroupIngressResponse_clusterSecurityGroup :: Lens' AuthorizeClusterSecurityGroupIngressResponse (Maybe ClusterSecurityGroup)

Undocumented member.

authorizeClusterSecurityGroupIngressResponse_httpStatus :: Lens' AuthorizeClusterSecurityGroupIngressResponse Int

The response's http status code.