Network.AWS.Redshift.AuthorizeEndpointAccess

Description

Grants access to a cluster.

Creating a Request

data AuthorizeEndpointAccess

See: newAuthorizeEndpointAccess smart constructor.

Constructors

AuthorizeEndpointAccess'

Fields

clusterIdentifier :: Maybe Text

The cluster identifier of the cluster to grant access to.

vpcIds :: Maybe [Text]

The virtual private cloud (VPC) identifiers to grant access to.

account :: Text

The Amazon Web Services account ID to grant access to.

newAuthorizeEndpointAccess

Arguments

:: Text
$sel:account:AuthorizeEndpointAccess'

-> AuthorizeEndpointAccess

Create a value of AuthorizeEndpointAccess with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterIdentifier:AuthorizeEndpointAccess', authorizeEndpointAccess_clusterIdentifier - The cluster identifier of the cluster to grant access to.

$sel:vpcIds:AuthorizeEndpointAccess', authorizeEndpointAccess_vpcIds - The virtual private cloud (VPC) identifiers to grant access to.

$sel:account:AuthorizeEndpointAccess', authorizeEndpointAccess_account - The Amazon Web Services account ID to grant access to.

Request Lenses

authorizeEndpointAccess_clusterIdentifier :: Lens' AuthorizeEndpointAccess (Maybe Text)

The cluster identifier of the cluster to grant access to.

authorizeEndpointAccess_vpcIds :: Lens' AuthorizeEndpointAccess (Maybe [Text])

The virtual private cloud (VPC) identifiers to grant access to.

authorizeEndpointAccess_account :: Lens' AuthorizeEndpointAccess Text

The Amazon Web Services account ID to grant access to.

Destructuring the Response

data EndpointAuthorization

Describes an endpoint authorization for authorizing Redshift-managed VPC endpoint access to a cluster across Amazon Web Services accounts.

See: newEndpointAuthorization smart constructor.

Constructors

EndpointAuthorization'

Fields

allowedAllVPCs :: Maybe Bool

Indicates whether all VPCs in the grantee account are allowed access to the cluster.

status :: Maybe AuthorizationStatus

The status of the authorization action.

clusterIdentifier :: Maybe Text

The cluster identifier.

grantee :: Maybe Text

The Amazon Web Services account ID of the grantee of the cluster.

authorizeTime :: Maybe ISO8601

The time (UTC) when the authorization was created.

allowedVPCs :: Maybe [Text]

The VPCs allowed access to the cluster.

clusterStatus :: Maybe Text

The status of the cluster.

grantor :: Maybe Text

The Amazon Web Services account ID of the cluster owner.

endpointCount :: Maybe Int

The number of Redshift-managed VPC endpoints created for the authorization.

newEndpointAuthorization :: EndpointAuthorization

Create a value of EndpointAuthorization with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:allowedAllVPCs:EndpointAuthorization', endpointAuthorization_allowedAllVPCs - Indicates whether all VPCs in the grantee account are allowed access to the cluster.

$sel:status:EndpointAuthorization', endpointAuthorization_status - The status of the authorization action.

$sel:clusterIdentifier:EndpointAuthorization', endpointAuthorization_clusterIdentifier - The cluster identifier.

$sel:grantee:EndpointAuthorization', endpointAuthorization_grantee - The Amazon Web Services account ID of the grantee of the cluster.

$sel:authorizeTime:EndpointAuthorization', endpointAuthorization_authorizeTime - The time (UTC) when the authorization was created.

$sel:allowedVPCs:EndpointAuthorization', endpointAuthorization_allowedVPCs - The VPCs allowed access to the cluster.

$sel:clusterStatus:EndpointAuthorization', endpointAuthorization_clusterStatus - The status of the cluster.

$sel:grantor:EndpointAuthorization', endpointAuthorization_grantor - The Amazon Web Services account ID of the cluster owner.

$sel:endpointCount:EndpointAuthorization', endpointAuthorization_endpointCount - The number of Redshift-managed VPC endpoints created for the authorization.

Response Lenses

endpointAuthorization_allowedAllVPCs :: Lens' EndpointAuthorization (Maybe Bool)

Indicates whether all VPCs in the grantee account are allowed access to the cluster.

endpointAuthorization_status :: Lens' EndpointAuthorization (Maybe AuthorizationStatus)

The status of the authorization action.

endpointAuthorization_clusterIdentifier :: Lens' EndpointAuthorization (Maybe Text)

The cluster identifier.

endpointAuthorization_grantee :: Lens' EndpointAuthorization (Maybe Text)

The Amazon Web Services account ID of the grantee of the cluster.

endpointAuthorization_authorizeTime :: Lens' EndpointAuthorization (Maybe UTCTime)

The time (UTC) when the authorization was created.

endpointAuthorization_allowedVPCs :: Lens' EndpointAuthorization (Maybe [Text])

The VPCs allowed access to the cluster.

endpointAuthorization_clusterStatus :: Lens' EndpointAuthorization (Maybe Text)

The status of the cluster.

endpointAuthorization_grantor :: Lens' EndpointAuthorization (Maybe Text)

The Amazon Web Services account ID of the cluster owner.

endpointAuthorization_endpointCount :: Lens' EndpointAuthorization (Maybe Int)

The number of Redshift-managed VPC endpoints created for the authorization.