Network.AWS.Redshift.AuthorizeSnapshotAccess

Description

Authorizes the specified Amazon Web Services account to restore the specified snapshot.

For more information about working with snapshots, go to Amazon Redshift Snapshots in the Amazon Redshift Cluster Management Guide.

Creating a Request

data AuthorizeSnapshotAccess

See: newAuthorizeSnapshotAccess smart constructor.

Constructors

AuthorizeSnapshotAccess'

Fields

snapshotClusterIdentifier :: Maybe Text

The identifier of the cluster the snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

snapshotIdentifier :: Text

The identifier of the snapshot the account is authorized to restore.

accountWithRestoreAccess :: Text

The identifier of the Amazon Web Services account authorized to restore the specified snapshot.

To share a snapshot with Amazon Web Services Support, specify amazon-redshift-support.

newAuthorizeSnapshotAccess

Arguments

:: Text
$sel:snapshotIdentifier:AuthorizeSnapshotAccess'

-> Text
$sel:accountWithRestoreAccess:AuthorizeSnapshotAccess'

-> AuthorizeSnapshotAccess

Create a value of AuthorizeSnapshotAccess with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshotClusterIdentifier:AuthorizeSnapshotAccess', authorizeSnapshotAccess_snapshotClusterIdentifier - The identifier of the cluster the snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

$sel:snapshotIdentifier:AuthorizeSnapshotAccess', authorizeSnapshotAccess_snapshotIdentifier - The identifier of the snapshot the account is authorized to restore.

$sel:accountWithRestoreAccess:AuthorizeSnapshotAccess', authorizeSnapshotAccess_accountWithRestoreAccess - The identifier of the Amazon Web Services account authorized to restore the specified snapshot.

To share a snapshot with Amazon Web Services Support, specify amazon-redshift-support.

Request Lenses

authorizeSnapshotAccess_snapshotClusterIdentifier :: Lens' AuthorizeSnapshotAccess (Maybe Text)

The identifier of the cluster the snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

authorizeSnapshotAccess_snapshotIdentifier :: Lens' AuthorizeSnapshotAccess Text

The identifier of the snapshot the account is authorized to restore.

authorizeSnapshotAccess_accountWithRestoreAccess :: Lens' AuthorizeSnapshotAccess Text

The identifier of the Amazon Web Services account authorized to restore the specified snapshot.

To share a snapshot with Amazon Web Services Support, specify amazon-redshift-support.

Destructuring the Response

data AuthorizeSnapshotAccessResponse

See: newAuthorizeSnapshotAccessResponse smart constructor.

Constructors

AuthorizeSnapshotAccessResponse'

Fields

snapshot :: Maybe Snapshot

httpStatus :: Int

The response's http status code.

newAuthorizeSnapshotAccessResponse

Arguments

:: Int
$sel:httpStatus:AuthorizeSnapshotAccessResponse'

-> AuthorizeSnapshotAccessResponse

Create a value of AuthorizeSnapshotAccessResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshot:AuthorizeSnapshotAccessResponse', authorizeSnapshotAccessResponse_snapshot - Undocumented member.

$sel:httpStatus:AuthorizeSnapshotAccessResponse', authorizeSnapshotAccessResponse_httpStatus - The response's http status code.

Response Lenses

authorizeSnapshotAccessResponse_snapshot :: Lens' AuthorizeSnapshotAccessResponse (Maybe Snapshot)

Undocumented member.

authorizeSnapshotAccessResponse_httpStatus :: Lens' AuthorizeSnapshotAccessResponse Int

The response's http status code.