Network.AWS.Redshift.CopyClusterSnapshot

Description

Copies the specified automated cluster snapshot to a new manual cluster snapshot. The source must be an automated snapshot and it must be in the available state.

When you delete a cluster, Amazon Redshift deletes any automated snapshots of the cluster. Also, when the retention period of the snapshot expires, Amazon Redshift automatically deletes it. If you want to keep an automated snapshot for a longer period, you can make a manual copy of the snapshot. Manual snapshots are retained until you delete them.

For more information about working with snapshots, go to Amazon Redshift Snapshots in the Amazon Redshift Cluster Management Guide.

Creating a Request

data CopyClusterSnapshot

See: newCopyClusterSnapshot smart constructor.

Constructors

CopyClusterSnapshot'

Fields

manualSnapshotRetentionPeriod :: Maybe Int

The number of days that a manual snapshot is retained. If the value is -1, the manual snapshot is retained indefinitely.

The value must be either -1 or an integer between 1 and 3,653.

The default value is -1.

sourceSnapshotClusterIdentifier :: Maybe Text

The identifier of the cluster the source snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

Constraints:

Must be the identifier for a valid cluster.

sourceSnapshotIdentifier :: Text

The identifier for the source snapshot.

Constraints:

Must be the identifier for a valid automated snapshot whose state is available.

targetSnapshotIdentifier :: Text

The identifier given to the new manual snapshot.

Constraints:

Cannot be null, empty, or blank.

Must contain from 1 to 255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Must be unique for the Amazon Web Services account that is making the request.

newCopyClusterSnapshot

Arguments

:: Text
$sel:sourceSnapshotIdentifier:CopyClusterSnapshot'

-> Text
$sel:targetSnapshotIdentifier:CopyClusterSnapshot'

-> CopyClusterSnapshot

Create a value of CopyClusterSnapshot with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:manualSnapshotRetentionPeriod:CopyClusterSnapshot', copyClusterSnapshot_manualSnapshotRetentionPeriod - The number of days that a manual snapshot is retained. If the value is -1, the manual snapshot is retained indefinitely.

The value must be either -1 or an integer between 1 and 3,653.

The default value is -1.

$sel:sourceSnapshotClusterIdentifier:CopyClusterSnapshot', copyClusterSnapshot_sourceSnapshotClusterIdentifier - The identifier of the cluster the source snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

Constraints:

Must be the identifier for a valid cluster.

$sel:sourceSnapshotIdentifier:CopyClusterSnapshot', copyClusterSnapshot_sourceSnapshotIdentifier - The identifier for the source snapshot.

Constraints:

Must be the identifier for a valid automated snapshot whose state is available.

$sel:targetSnapshotIdentifier:CopyClusterSnapshot', copyClusterSnapshot_targetSnapshotIdentifier - The identifier given to the new manual snapshot.

Constraints:

Cannot be null, empty, or blank.

Must contain from 1 to 255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Must be unique for the Amazon Web Services account that is making the request.

Request Lenses

copyClusterSnapshot_manualSnapshotRetentionPeriod :: Lens' CopyClusterSnapshot (Maybe Int)

The number of days that a manual snapshot is retained. If the value is -1, the manual snapshot is retained indefinitely.

The value must be either -1 or an integer between 1 and 3,653.

The default value is -1.

copyClusterSnapshot_sourceSnapshotClusterIdentifier :: Lens' CopyClusterSnapshot (Maybe Text)

The identifier of the cluster the source snapshot was created from. This parameter is required if your IAM user has a policy containing a snapshot resource element that specifies anything other than * for the cluster name.

Constraints:

Must be the identifier for a valid cluster.

copyClusterSnapshot_sourceSnapshotIdentifier :: Lens' CopyClusterSnapshot Text

The identifier for the source snapshot.

Constraints:

Must be the identifier for a valid automated snapshot whose state is available.

copyClusterSnapshot_targetSnapshotIdentifier :: Lens' CopyClusterSnapshot Text

The identifier given to the new manual snapshot.

Constraints:

Cannot be null, empty, or blank.

Must contain from 1 to 255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Must be unique for the Amazon Web Services account that is making the request.

Destructuring the Response

data CopyClusterSnapshotResponse

See: newCopyClusterSnapshotResponse smart constructor.

Constructors

CopyClusterSnapshotResponse'

Fields

snapshot :: Maybe Snapshot

httpStatus :: Int

The response's http status code.

newCopyClusterSnapshotResponse

Arguments

:: Int
$sel:httpStatus:CopyClusterSnapshotResponse'

-> CopyClusterSnapshotResponse

Create a value of CopyClusterSnapshotResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshot:CopyClusterSnapshotResponse', copyClusterSnapshotResponse_snapshot - Undocumented member.

$sel:httpStatus:CopyClusterSnapshotResponse', copyClusterSnapshotResponse_httpStatus - The response's http status code.

Response Lenses

copyClusterSnapshotResponse_snapshot :: Lens' CopyClusterSnapshotResponse (Maybe Snapshot)

Undocumented member.

copyClusterSnapshotResponse_httpStatus :: Lens' CopyClusterSnapshotResponse Int

The response's http status code.