Network.AWS.Redshift.CreateCluster

Description

Creates a new cluster with the specified parameters.

To create a cluster in Virtual Private Cloud (VPC), you must provide a cluster subnet group name. The cluster subnet group identifies the subnets of your VPC that Amazon Redshift uses when creating the cluster. For more information about managing clusters, go to Amazon Redshift Clusters in the Amazon Redshift Cluster Management Guide.

Creating a Request

data CreateCluster

See: newCreateCluster smart constructor.

Constructors

CreateCluster'

Fields

enhancedVpcRouting :: Maybe Bool

An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide.

If this option is true, enhanced VPC routing is enabled.

Default: false

additionalInfo :: Maybe Text

Reserved.

elasticIp :: Maybe Text

The Elastic IP (EIP) address for the cluster.

Constraints: The cluster must be provisioned in EC2-VPC and publicly-accessible through an Internet gateway. For more information about provisioning clusters in EC2-VPC, go to Supported Platforms to Launch Your Cluster in the Amazon Redshift Cluster Management Guide.

clusterSubnetGroupName :: Maybe Text

The name of a cluster subnet group to be associated with this cluster.

If this parameter is not provided the resulting cluster will be deployed outside virtual private cloud (VPC).

hsmClientCertificateIdentifier :: Maybe Text

Specifies the name of the HSM client certificate the Amazon Redshift cluster uses to retrieve the data encryption keys stored in an HSM.

encrypted :: Maybe Bool

If true, the data in the cluster is encrypted at rest.

Default: false

allowVersionUpgrade :: Maybe Bool

If true, major version upgrades can be applied during the maintenance window to the Amazon Redshift engine that is running on the cluster.

When a new major version of the Amazon Redshift engine is released, you can request that the service automatically apply upgrades during the maintenance window to the Amazon Redshift engine that is running on your cluster.

Default: true

clusterParameterGroupName :: Maybe Text

The name of the parameter group to be associated with this cluster.

Default: The default Amazon Redshift cluster parameter group. For information about the default parameter group, go to Working with Amazon Redshift Parameter Groups

Constraints:

Must be 1 to 255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

automatedSnapshotRetentionPeriod :: Maybe Int

The number of days that automated snapshots are retained. If the value is 0, automated snapshots are disabled. Even if automated snapshots are disabled, you can still create manual snapshots when you want with CreateClusterSnapshot.

You can't disable automated snapshots for RA3 node types. Set the automated retention period from 1-35 days.

Default: 1

Constraints: Must be a value from 0 to 35.

availabilityZoneRelocation :: Maybe Bool

The option to enable relocation for an Amazon Redshift cluster between Availability Zones after the cluster is created.

publiclyAccessible :: Maybe Bool

If true, the cluster can be accessed from a public network.

clusterType :: Maybe Text

The type of the cluster. When cluster type is specified as

single-node, the NumberOfNodes parameter is not required.

multi-node, the NumberOfNodes parameter is required.

Valid Values: multi-node | single-node

Default: multi-node

snapshotScheduleIdentifier :: Maybe Text

A unique identifier for the snapshot schedule.

vpcSecurityGroupIds :: Maybe [Text]

A list of Virtual Private Cloud (VPC) security groups to be associated with the cluster.

Default: The default VPC security group is associated with the cluster.

kmsKeyId :: Maybe Text

The Key Management Service (KMS) key ID of the encryption key that you want to use to encrypt data in the cluster.

manualSnapshotRetentionPeriod :: Maybe Int

The default number of days to retain a manual snapshot. If the value is -1, the snapshot is retained indefinitely. This setting doesn't change the retention period of existing snapshots.

The value must be either -1 or an integer between 1 and 3,653.

availabilityZone :: Maybe Text

The EC2 Availability Zone (AZ) in which you want Amazon Redshift to provision the cluster. For example, if you have several EC2 instances running in a specific Availability Zone, then you might want the cluster to be provisioned in the same zone in order to decrease network latency.

Default: A random, system-chosen Availability Zone in the region that is specified by the endpoint.

Example: us-east-2d

Constraint: The specified Availability Zone must be in the same region as the current endpoint.

preferredMaintenanceWindow :: Maybe Text

The weekly time range (in UTC) during which automated cluster maintenance can occur.

Format: ddd:hh24:mi-ddd:hh24:mi

Default: A 30-minute window selected at random from an 8-hour block of time per region, occurring on a random day of the week. For more information about the time blocks for each region, see Maintenance Windows in Amazon Redshift Cluster Management Guide.

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Minimum 30-minute window.

tags :: Maybe [Tag]

A list of tag instances.

dbName :: Maybe Text

The name of the first database to be created when the cluster is created.

To create additional databases after the cluster is created, connect to the cluster with a SQL client and use SQL commands to create a database. For more information, go to Create a Database in the Amazon Redshift Database Developer Guide.

Default: dev

Constraints:

Must contain 1 to 64 alphanumeric characters.

Must contain only lowercase letters.

Cannot be a word that is reserved by the service. A list of reserved words can be found in Reserved Words in the Amazon Redshift Database Developer Guide.

port :: Maybe Int

The port number on which the cluster accepts incoming connections.

The cluster is accessible only via the JDBC and ODBC connection strings. Part of the connection string requires the port on which the cluster will listen for incoming connections.

Default: 5439

Valid Values: 1150-65535

numberOfNodes :: Maybe Int

The number of compute nodes in the cluster. This parameter is required when the ClusterType parameter is specified as multi-node.

For information about determining how many nodes you need, go to Working with Clusters in the Amazon Redshift Cluster Management Guide.

If you don't specify this parameter, you get a single-node cluster. When requesting a multi-node cluster, you must specify the number of nodes that you want in the cluster.

Default: 1

Constraints: Value must be at least 1 and no more than 100.

clusterVersion :: Maybe Text

The version of the Amazon Redshift engine software that you want to deploy on the cluster.

The version selected runs on all the nodes in the cluster.

Constraints: Only version 1.0 is currently available.

Example: 1.0

clusterSecurityGroups :: Maybe [Text]

A list of security groups to be associated with this cluster.

Default: The default cluster security group for Amazon Redshift.

aquaConfigurationStatus :: Maybe AquaConfigurationStatus

The value represents how the cluster is configured to use AQUA (Advanced Query Accelerator) when it is created. Possible values include the following.

enabled - Use AQUA if it is available for the current Amazon Web Services Region and Amazon Redshift node type.

disabled - Don't use AQUA.

auto - Amazon Redshift determines whether to use AQUA.

hsmConfigurationIdentifier :: Maybe Text

Specifies the name of the HSM configuration that contains the information the Amazon Redshift cluster can use to retrieve and store keys in an HSM.

maintenanceTrackName :: Maybe Text

An optional parameter for the name of the maintenance track for the cluster. If you don't provide a maintenance track name, the cluster is assigned to the current track.

iamRoles :: Maybe [Text]

A list of Identity and Access Management (IAM) roles that can be used by the cluster to access other Amazon Web Services services. You must supply the IAM roles in their Amazon Resource Name (ARN) format. You can supply up to 10 IAM roles in a single request.

A cluster can have up to 10 IAM roles associated with it at any time.

clusterIdentifier :: Text

A unique identifier for the cluster. You use this identifier to refer to the cluster for any subsequent cluster operations such as deleting or modifying. The identifier also appears in the Amazon Redshift console.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens.

Alphabetic characters must be lowercase.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Must be unique for all clusters within an Amazon Web Services account.

Example: myexamplecluster

nodeType :: Text

The node type to be provisioned for the cluster. For information about node types, go to Working with Clusters in the Amazon Redshift Cluster Management Guide.

Valid Values: ds2.xlarge | ds2.8xlarge | dc1.large | dc1.8xlarge | dc2.large | dc2.8xlarge | ra3.xlplus | ra3.4xlarge | ra3.16xlarge

masterUsername :: Text

The user name associated with the admin user account for the cluster that is being created.

Constraints:

Must be 1 - 128 alphanumeric characters. The user name can't be PUBLIC.

First character must be a letter.

Cannot be a reserved word. A list of reserved words can be found in Reserved Words in the Amazon Redshift Database Developer Guide.

masterUserPassword :: Text

The password associated with the admin user account for the cluster that is being created.

Constraints:

Must be between 8 and 64 characters in length.

Must contain at least one uppercase letter.

Must contain at least one lowercase letter.

Must contain one number.

Can be any printable ASCII character (ASCII code 33 to 126) except ' (single quote), " (double quote), \, /, @, or space.

newCreateCluster

Arguments

:: Text
$sel:clusterIdentifier:CreateCluster'

-> Text
$sel:nodeType:CreateCluster'

-> Text
$sel:masterUsername:CreateCluster'

-> Text
$sel:masterUserPassword:CreateCluster'

-> CreateCluster

Create a value of CreateCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enhancedVpcRouting:CreateCluster', createCluster_enhancedVpcRouting - An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide.

If this option is true, enhanced VPC routing is enabled.

Default: false

$sel:additionalInfo:CreateCluster', createCluster_additionalInfo - Reserved.

$sel:elasticIp:CreateCluster', createCluster_elasticIp - The Elastic IP (EIP) address for the cluster.

Constraints: The cluster must be provisioned in EC2-VPC and publicly-accessible through an Internet gateway. For more information about provisioning clusters in EC2-VPC, go to Supported Platforms to Launch Your Cluster in the Amazon Redshift Cluster Management Guide.

$sel:clusterSubnetGroupName:CreateCluster', createCluster_clusterSubnetGroupName - The name of a cluster subnet group to be associated with this cluster.

If this parameter is not provided the resulting cluster will be deployed outside virtual private cloud (VPC).

$sel:hsmClientCertificateIdentifier:CreateCluster', createCluster_hsmClientCertificateIdentifier - Specifies the name of the HSM client certificate the Amazon Redshift cluster uses to retrieve the data encryption keys stored in an HSM.

$sel:encrypted:CreateCluster', createCluster_encrypted - If true, the data in the cluster is encrypted at rest.

Default: false

$sel:allowVersionUpgrade:CreateCluster', createCluster_allowVersionUpgrade - If true, major version upgrades can be applied during the maintenance window to the Amazon Redshift engine that is running on the cluster.

When a new major version of the Amazon Redshift engine is released, you can request that the service automatically apply upgrades during the maintenance window to the Amazon Redshift engine that is running on your cluster.

Default: true

$sel:clusterParameterGroupName:CreateCluster', createCluster_clusterParameterGroupName - The name of the parameter group to be associated with this cluster.

Default: The default Amazon Redshift cluster parameter group. For information about the default parameter group, go to Working with Amazon Redshift Parameter Groups

Constraints:

Must be 1 to 255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

$sel:automatedSnapshotRetentionPeriod:CreateCluster', createCluster_automatedSnapshotRetentionPeriod - The number of days that automated snapshots are retained. If the value is 0, automated snapshots are disabled. Even if automated snapshots are disabled, you can still create manual snapshots when you want with CreateClusterSnapshot.

You can't disable automated snapshots for RA3 node types. Set the automated retention period from 1-35 days.

Default: 1

Constraints: Must be a value from 0 to 35.

$sel:availabilityZoneRelocation:CreateCluster', createCluster_availabilityZoneRelocation - The option to enable relocation for an Amazon Redshift cluster between Availability Zones after the cluster is created.

$sel:publiclyAccessible:CreateCluster', createCluster_publiclyAccessible - If true, the cluster can be accessed from a public network.

$sel:clusterType:CreateCluster', createCluster_clusterType - The type of the cluster. When cluster type is specified as

single-node, the NumberOfNodes parameter is not required.

multi-node, the NumberOfNodes parameter is required.

Valid Values: multi-node | single-node

Default: multi-node

$sel:snapshotScheduleIdentifier:CreateCluster', createCluster_snapshotScheduleIdentifier - A unique identifier for the snapshot schedule.

$sel:vpcSecurityGroupIds:CreateCluster', createCluster_vpcSecurityGroupIds - A list of Virtual Private Cloud (VPC) security groups to be associated with the cluster.

Default: The default VPC security group is associated with the cluster.

$sel:kmsKeyId:CreateCluster', createCluster_kmsKeyId - The Key Management Service (KMS) key ID of the encryption key that you want to use to encrypt data in the cluster.

$sel:manualSnapshotRetentionPeriod:CreateCluster', createCluster_manualSnapshotRetentionPeriod - The default number of days to retain a manual snapshot. If the value is -1, the snapshot is retained indefinitely. This setting doesn't change the retention period of existing snapshots.

The value must be either -1 or an integer between 1 and 3,653.

$sel:availabilityZone:CreateCluster', createCluster_availabilityZone - The EC2 Availability Zone (AZ) in which you want Amazon Redshift to provision the cluster. For example, if you have several EC2 instances running in a specific Availability Zone, then you might want the cluster to be provisioned in the same zone in order to decrease network latency.

Default: A random, system-chosen Availability Zone in the region that is specified by the endpoint.

Example: us-east-2d

Constraint: The specified Availability Zone must be in the same region as the current endpoint.

$sel:preferredMaintenanceWindow:CreateCluster', createCluster_preferredMaintenanceWindow - The weekly time range (in UTC) during which automated cluster maintenance can occur.

Format: ddd:hh24:mi-ddd:hh24:mi

Default: A 30-minute window selected at random from an 8-hour block of time per region, occurring on a random day of the week. For more information about the time blocks for each region, see Maintenance Windows in Amazon Redshift Cluster Management Guide.

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Minimum 30-minute window.

$sel:tags:CreateCluster', createCluster_tags - A list of tag instances.

$sel:dbName:CreateCluster', createCluster_dbName - The name of the first database to be created when the cluster is created.

To create additional databases after the cluster is created, connect to the cluster with a SQL client and use SQL commands to create a database. For more information, go to Create a Database in the Amazon Redshift Database Developer Guide.

Default: dev

Constraints:

Must contain 1 to 64 alphanumeric characters.

Must contain only lowercase letters.

Cannot be a word that is reserved by the service. A list of reserved words can be found in Reserved Words in the Amazon Redshift Database Developer Guide.

$sel:port:CreateCluster', createCluster_port - The port number on which the cluster accepts incoming connections.

The cluster is accessible only via the JDBC and ODBC connection strings. Part of the connection string requires the port on which the cluster will listen for incoming connections.

Default: 5439

Valid Values: 1150-65535

$sel:numberOfNodes:CreateCluster', createCluster_numberOfNodes - The number of compute nodes in the cluster. This parameter is required when the ClusterType parameter is specified as multi-node.

For information about determining how many nodes you need, go to Working with Clusters in the Amazon Redshift Cluster Management Guide.

If you don't specify this parameter, you get a single-node cluster. When requesting a multi-node cluster, you must specify the number of nodes that you want in the cluster.

Default: 1

Constraints: Value must be at least 1 and no more than 100.

$sel:clusterVersion:CreateCluster', createCluster_clusterVersion - The version of the Amazon Redshift engine software that you want to deploy on the cluster.

The version selected runs on all the nodes in the cluster.

Constraints: Only version 1.0 is currently available.

Example: 1.0

$sel:clusterSecurityGroups:CreateCluster', createCluster_clusterSecurityGroups - A list of security groups to be associated with this cluster.

Default: The default cluster security group for Amazon Redshift.

$sel:aquaConfigurationStatus:CreateCluster', createCluster_aquaConfigurationStatus - The value represents how the cluster is configured to use AQUA (Advanced Query Accelerator) when it is created. Possible values include the following.

enabled - Use AQUA if it is available for the current Amazon Web Services Region and Amazon Redshift node type.

disabled - Don't use AQUA.

auto - Amazon Redshift determines whether to use AQUA.

$sel:hsmConfigurationIdentifier:CreateCluster', createCluster_hsmConfigurationIdentifier - Specifies the name of the HSM configuration that contains the information the Amazon Redshift cluster can use to retrieve and store keys in an HSM.

$sel:maintenanceTrackName:CreateCluster', createCluster_maintenanceTrackName - An optional parameter for the name of the maintenance track for the cluster. If you don't provide a maintenance track name, the cluster is assigned to the current track.

$sel:iamRoles:CreateCluster', createCluster_iamRoles - A list of Identity and Access Management (IAM) roles that can be used by the cluster to access other Amazon Web Services services. You must supply the IAM roles in their Amazon Resource Name (ARN) format. You can supply up to 10 IAM roles in a single request.

A cluster can have up to 10 IAM roles associated with it at any time.

$sel:clusterIdentifier:CreateCluster', createCluster_clusterIdentifier - A unique identifier for the cluster. You use this identifier to refer to the cluster for any subsequent cluster operations such as deleting or modifying. The identifier also appears in the Amazon Redshift console.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens.

Alphabetic characters must be lowercase.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Must be unique for all clusters within an Amazon Web Services account.

Example: myexamplecluster

$sel:nodeType:CreateCluster', createCluster_nodeType - The node type to be provisioned for the cluster. For information about node types, go to Working with Clusters in the Amazon Redshift Cluster Management Guide.

Valid Values: ds2.xlarge | ds2.8xlarge | dc1.large | dc1.8xlarge | dc2.large | dc2.8xlarge | ra3.xlplus | ra3.4xlarge | ra3.16xlarge

$sel:masterUsername:CreateCluster', createCluster_masterUsername - The user name associated with the admin user account for the cluster that is being created.

Constraints:

Must be 1 - 128 alphanumeric characters. The user name can't be PUBLIC.

First character must be a letter.

Cannot be a reserved word. A list of reserved words can be found in Reserved Words in the Amazon Redshift Database Developer Guide.

$sel:masterUserPassword:CreateCluster', createCluster_masterUserPassword - The password associated with the admin user account for the cluster that is being created.

Constraints:

Must be between 8 and 64 characters in length.

Must contain at least one uppercase letter.

Must contain at least one lowercase letter.

Must contain one number.

Can be any printable ASCII character (ASCII code 33 to 126) except ' (single quote), " (double quote), \, /, @, or space.

Request Lenses

createCluster_enhancedVpcRouting :: Lens' CreateCluster (Maybe Bool)

An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide.

If this option is true, enhanced VPC routing is enabled.

Default: false

createCluster_additionalInfo :: Lens' CreateCluster (Maybe Text)

Reserved.

createCluster_elasticIp :: Lens' CreateCluster (Maybe Text)

The Elastic IP (EIP) address for the cluster.

Constraints: The cluster must be provisioned in EC2-VPC and publicly-accessible through an Internet gateway. For more information about provisioning clusters in EC2-VPC, go to Supported Platforms to Launch Your Cluster in the Amazon Redshift Cluster Management Guide.

createCluster_clusterSubnetGroupName :: Lens' CreateCluster (Maybe Text)

The name of a cluster subnet group to be associated with this cluster.

If this parameter is not provided the resulting cluster will be deployed outside virtual private cloud (VPC).

createCluster_hsmClientCertificateIdentifier :: Lens' CreateCluster (Maybe Text)

Specifies the name of the HSM client certificate the Amazon Redshift cluster uses to retrieve the data encryption keys stored in an HSM.

createCluster_encrypted :: Lens' CreateCluster (Maybe Bool)

If true, the data in the cluster is encrypted at rest.

Default: false

createCluster_allowVersionUpgrade :: Lens' CreateCluster (Maybe Bool)

If true, major version upgrades can be applied during the maintenance window to the Amazon Redshift engine that is running on the cluster.

When a new major version of the Amazon Redshift engine is released, you can request that the service automatically apply upgrades during the maintenance window to the Amazon Redshift engine that is running on your cluster.

Default: true

createCluster_clusterParameterGroupName :: Lens' CreateCluster (Maybe Text)

The name of the parameter group to be associated with this cluster.

Default: The default Amazon Redshift cluster parameter group. For information about the default parameter group, go to Working with Amazon Redshift Parameter Groups

Constraints:

Must be 1 to 255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

createCluster_automatedSnapshotRetentionPeriod :: Lens' CreateCluster (Maybe Int)

The number of days that automated snapshots are retained. If the value is 0, automated snapshots are disabled. Even if automated snapshots are disabled, you can still create manual snapshots when you want with CreateClusterSnapshot.

You can't disable automated snapshots for RA3 node types. Set the automated retention period from 1-35 days.

Default: 1

Constraints: Must be a value from 0 to 35.

createCluster_availabilityZoneRelocation :: Lens' CreateCluster (Maybe Bool)

The option to enable relocation for an Amazon Redshift cluster between Availability Zones after the cluster is created.

createCluster_publiclyAccessible :: Lens' CreateCluster (Maybe Bool)

If true, the cluster can be accessed from a public network.

createCluster_clusterType :: Lens' CreateCluster (Maybe Text)

The type of the cluster. When cluster type is specified as

single-node, the NumberOfNodes parameter is not required.

multi-node, the NumberOfNodes parameter is required.

Valid Values: multi-node | single-node

Default: multi-node

createCluster_snapshotScheduleIdentifier :: Lens' CreateCluster (Maybe Text)

A unique identifier for the snapshot schedule.

createCluster_vpcSecurityGroupIds :: Lens' CreateCluster (Maybe [Text])

A list of Virtual Private Cloud (VPC) security groups to be associated with the cluster.

Default: The default VPC security group is associated with the cluster.

createCluster_kmsKeyId :: Lens' CreateCluster (Maybe Text)

The Key Management Service (KMS) key ID of the encryption key that you want to use to encrypt data in the cluster.

createCluster_manualSnapshotRetentionPeriod :: Lens' CreateCluster (Maybe Int)

The default number of days to retain a manual snapshot. If the value is -1, the snapshot is retained indefinitely. This setting doesn't change the retention period of existing snapshots.

The value must be either -1 or an integer between 1 and 3,653.

createCluster_availabilityZone :: Lens' CreateCluster (Maybe Text)

The EC2 Availability Zone (AZ) in which you want Amazon Redshift to provision the cluster. For example, if you have several EC2 instances running in a specific Availability Zone, then you might want the cluster to be provisioned in the same zone in order to decrease network latency.

Default: A random, system-chosen Availability Zone in the region that is specified by the endpoint.

Example: us-east-2d

Constraint: The specified Availability Zone must be in the same region as the current endpoint.

createCluster_preferredMaintenanceWindow :: Lens' CreateCluster (Maybe Text)

The weekly time range (in UTC) during which automated cluster maintenance can occur.

Format: ddd:hh24:mi-ddd:hh24:mi

Default: A 30-minute window selected at random from an 8-hour block of time per region, occurring on a random day of the week. For more information about the time blocks for each region, see Maintenance Windows in Amazon Redshift Cluster Management Guide.

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Minimum 30-minute window.

createCluster_tags :: Lens' CreateCluster (Maybe [Tag])

A list of tag instances.

createCluster_dbName :: Lens' CreateCluster (Maybe Text)

The name of the first database to be created when the cluster is created.

To create additional databases after the cluster is created, connect to the cluster with a SQL client and use SQL commands to create a database. For more information, go to Create a Database in the Amazon Redshift Database Developer Guide.

Default: dev

Constraints:

Must contain 1 to 64 alphanumeric characters.

Must contain only lowercase letters.

Cannot be a word that is reserved by the service. A list of reserved words can be found in Reserved Words in the Amazon Redshift Database Developer Guide.

createCluster_port :: Lens' CreateCluster (Maybe Int)

The port number on which the cluster accepts incoming connections.

The cluster is accessible only via the JDBC and ODBC connection strings. Part of the connection string requires the port on which the cluster will listen for incoming connections.

Default: 5439

Valid Values: 1150-65535

createCluster_numberOfNodes :: Lens' CreateCluster (Maybe Int)

The number of compute nodes in the cluster. This parameter is required when the ClusterType parameter is specified as multi-node.

For information about determining how many nodes you need, go to Working with Clusters in the Amazon Redshift Cluster Management Guide.

If you don't specify this parameter, you get a single-node cluster. When requesting a multi-node cluster, you must specify the number of nodes that you want in the cluster.

Default: 1

Constraints: Value must be at least 1 and no more than 100.

createCluster_clusterVersion :: Lens' CreateCluster (Maybe Text)

The version of the Amazon Redshift engine software that you want to deploy on the cluster.

The version selected runs on all the nodes in the cluster.

Constraints: Only version 1.0 is currently available.

Example: 1.0

createCluster_clusterSecurityGroups :: Lens' CreateCluster (Maybe [Text])

A list of security groups to be associated with this cluster.

Default: The default cluster security group for Amazon Redshift.

createCluster_aquaConfigurationStatus :: Lens' CreateCluster (Maybe AquaConfigurationStatus)

The value represents how the cluster is configured to use AQUA (Advanced Query Accelerator) when it is created. Possible values include the following.

enabled - Use AQUA if it is available for the current Amazon Web Services Region and Amazon Redshift node type.

disabled - Don't use AQUA.

auto - Amazon Redshift determines whether to use AQUA.

createCluster_hsmConfigurationIdentifier :: Lens' CreateCluster (Maybe Text)

Specifies the name of the HSM configuration that contains the information the Amazon Redshift cluster can use to retrieve and store keys in an HSM.

createCluster_maintenanceTrackName :: Lens' CreateCluster (Maybe Text)

An optional parameter for the name of the maintenance track for the cluster. If you don't provide a maintenance track name, the cluster is assigned to the current track.

createCluster_iamRoles :: Lens' CreateCluster (Maybe [Text])

A list of Identity and Access Management (IAM) roles that can be used by the cluster to access other Amazon Web Services services. You must supply the IAM roles in their Amazon Resource Name (ARN) format. You can supply up to 10 IAM roles in a single request.

A cluster can have up to 10 IAM roles associated with it at any time.

createCluster_clusterIdentifier :: Lens' CreateCluster Text

A unique identifier for the cluster. You use this identifier to refer to the cluster for any subsequent cluster operations such as deleting or modifying. The identifier also appears in the Amazon Redshift console.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens.

Alphabetic characters must be lowercase.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Must be unique for all clusters within an Amazon Web Services account.

Example: myexamplecluster

createCluster_nodeType :: Lens' CreateCluster Text

The node type to be provisioned for the cluster. For information about node types, go to Working with Clusters in the Amazon Redshift Cluster Management Guide.

Valid Values: ds2.xlarge | ds2.8xlarge | dc1.large | dc1.8xlarge | dc2.large | dc2.8xlarge | ra3.xlplus | ra3.4xlarge | ra3.16xlarge

createCluster_masterUsername :: Lens' CreateCluster Text

The user name associated with the admin user account for the cluster that is being created.

Constraints:

Must be 1 - 128 alphanumeric characters. The user name can't be PUBLIC.

First character must be a letter.

Cannot be a reserved word. A list of reserved words can be found in Reserved Words in the Amazon Redshift Database Developer Guide.

createCluster_masterUserPassword :: Lens' CreateCluster Text

The password associated with the admin user account for the cluster that is being created.

Constraints:

Must be between 8 and 64 characters in length.

Must contain at least one uppercase letter.

Must contain at least one lowercase letter.

Must contain one number.

Can be any printable ASCII character (ASCII code 33 to 126) except ' (single quote), " (double quote), \, /, @, or space.

Destructuring the Response

data CreateClusterResponse

See: newCreateClusterResponse smart constructor.

Constructors

CreateClusterResponse'

Fields

cluster :: Maybe Cluster

httpStatus :: Int

The response's http status code.

newCreateClusterResponse

Arguments

:: Int
$sel:httpStatus:CreateClusterResponse'

-> CreateClusterResponse

Create a value of CreateClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:CreateClusterResponse', createClusterResponse_cluster - Undocumented member.

$sel:httpStatus:CreateClusterResponse', createClusterResponse_httpStatus - The response's http status code.

Response Lenses

createClusterResponse_cluster :: Lens' CreateClusterResponse (Maybe Cluster)

Undocumented member.

createClusterResponse_httpStatus :: Lens' CreateClusterResponse Int

The response's http status code.