Network.AWS.Redshift.CreateClusterSecurityGroup

Description

Creates a new Amazon Redshift security group. You use security groups to control access to non-VPC clusters.

For information about managing security groups, go to Amazon Redshift Cluster Security Groups in the Amazon Redshift Cluster Management Guide.

Creating a Request

data CreateClusterSecurityGroup

See: newCreateClusterSecurityGroup smart constructor.

Constructors

CreateClusterSecurityGroup'

Fields

tags :: Maybe [Tag]

A list of tag instances.

clusterSecurityGroupName :: Text

The name for the security group. Amazon Redshift stores the value as a lowercase string.

Constraints:

Must contain no more than 255 alphanumeric characters or hyphens.

Must not be "Default".

Must be unique for all security groups that are created by your Amazon Web Services account.

Example: examplesecuritygroup

description :: Text

A description for the security group.

newCreateClusterSecurityGroup

Arguments

:: Text
$sel:clusterSecurityGroupName:CreateClusterSecurityGroup'

-> Text
$sel:description:CreateClusterSecurityGroup'

-> CreateClusterSecurityGroup

Create a value of CreateClusterSecurityGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateClusterSecurityGroup', createClusterSecurityGroup_tags - A list of tag instances.

$sel:clusterSecurityGroupName:CreateClusterSecurityGroup', createClusterSecurityGroup_clusterSecurityGroupName - The name for the security group. Amazon Redshift stores the value as a lowercase string.

Constraints:

Must contain no more than 255 alphanumeric characters or hyphens.

Must not be "Default".

Must be unique for all security groups that are created by your Amazon Web Services account.

Example: examplesecuritygroup

$sel:description:CreateClusterSecurityGroup', createClusterSecurityGroup_description - A description for the security group.

Request Lenses

createClusterSecurityGroup_tags :: Lens' CreateClusterSecurityGroup (Maybe [Tag])

A list of tag instances.

createClusterSecurityGroup_clusterSecurityGroupName :: Lens' CreateClusterSecurityGroup Text

The name for the security group. Amazon Redshift stores the value as a lowercase string.

Constraints:

Must contain no more than 255 alphanumeric characters or hyphens.

Must not be "Default".

Must be unique for all security groups that are created by your Amazon Web Services account.

Example: examplesecuritygroup

createClusterSecurityGroup_description :: Lens' CreateClusterSecurityGroup Text

A description for the security group.

Destructuring the Response

data CreateClusterSecurityGroupResponse

See: newCreateClusterSecurityGroupResponse smart constructor.

Constructors

CreateClusterSecurityGroupResponse'

Fields

clusterSecurityGroup :: Maybe ClusterSecurityGroup

httpStatus :: Int

The response's http status code.

newCreateClusterSecurityGroupResponse

Arguments

:: Int
$sel:httpStatus:CreateClusterSecurityGroupResponse'

-> CreateClusterSecurityGroupResponse

Create a value of CreateClusterSecurityGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterSecurityGroup:CreateClusterSecurityGroupResponse', createClusterSecurityGroupResponse_clusterSecurityGroup - Undocumented member.

$sel:httpStatus:CreateClusterSecurityGroupResponse', createClusterSecurityGroupResponse_httpStatus - The response's http status code.

Response Lenses

createClusterSecurityGroupResponse_clusterSecurityGroup :: Lens' CreateClusterSecurityGroupResponse (Maybe ClusterSecurityGroup)

Undocumented member.

createClusterSecurityGroupResponse_httpStatus :: Lens' CreateClusterSecurityGroupResponse Int

The response's http status code.