Network.AWS.Redshift.CreateEventSubscription

Description

Creates an Amazon Redshift event notification subscription. This action requires an ARN (Amazon Resource Name) of an Amazon SNS topic created by either the Amazon Redshift console, the Amazon SNS console, or the Amazon SNS API. To obtain an ARN with Amazon SNS, you must create a topic in Amazon SNS and subscribe to the topic. The ARN is displayed in the SNS console.

You can specify the source type, and lists of Amazon Redshift source IDs, event categories, and event severities. Notifications will be sent for all events you want that match those criteria. For example, you can specify source type = cluster, source ID = my-cluster-1 and mycluster2, event categories = Availability, Backup, and severity = ERROR. The subscription will only send notifications for those ERROR events in the Availability and Backup categories for the specified clusters.

If you specify both the source type and source IDs, such as source type = cluster and source identifier = my-cluster-1, notifications will be sent for all the cluster events for my-cluster-1. If you specify a source type but do not specify a source identifier, you will receive notice of the events for the objects of that type in your Amazon Web Services account. If you do not specify either the SourceType nor the SourceIdentifier, you will be notified of events generated from all Amazon Redshift sources belonging to your Amazon Web Services account. You must specify a source type if you specify a source ID.

Creating a Request

data CreateEventSubscription

See: newCreateEventSubscription smart constructor.

Constructors

CreateEventSubscription'

Fields

sourceIds :: Maybe [Text]

A list of one or more identifiers of Amazon Redshift source objects. All of the objects must be of the same type as was specified in the source type parameter. The event subscription will return only events generated by the specified objects. If not specified, then events are returned for all objects within the source type specified.

Example: my-cluster-1, my-cluster-2

Example: my-snapshot-20131010

severity :: Maybe Text

Specifies the Amazon Redshift event severity to be published by the event notification subscription.

Values: ERROR, INFO

enabled :: Maybe Bool

A boolean value; set to true to activate the subscription, and set to false to create the subscription but not activate it.

eventCategories :: Maybe [Text]

Specifies the Amazon Redshift event categories to be published by the event notification subscription.

Values: configuration, management, monitoring, security, pending

tags :: Maybe [Tag]

A list of tag instances.

sourceType :: Maybe Text

The type of source that will be generating the events. For example, if you want to be notified of events generated by a cluster, you would set this parameter to cluster. If this value is not specified, events are returned for all Amazon Redshift objects in your Amazon Web Services account. You must specify a source type in order to specify source IDs.

Valid values: cluster, cluster-parameter-group, cluster-security-group, cluster-snapshot, and scheduled-action.

subscriptionName :: Text

The name of the event subscription to be created.

Constraints:

Cannot be null, empty, or blank.

Must contain from 1 to 255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

snsTopicArn :: Text

The Amazon Resource Name (ARN) of the Amazon SNS topic used to transmit the event notifications. The ARN is created by Amazon SNS when you create a topic and subscribe to it.

newCreateEventSubscription

Arguments

:: Text
$sel:subscriptionName:CreateEventSubscription'

-> Text
$sel:snsTopicArn:CreateEventSubscription'

-> CreateEventSubscription

Create a value of CreateEventSubscription with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:sourceIds:CreateEventSubscription', createEventSubscription_sourceIds - A list of one or more identifiers of Amazon Redshift source objects. All of the objects must be of the same type as was specified in the source type parameter. The event subscription will return only events generated by the specified objects. If not specified, then events are returned for all objects within the source type specified.

Example: my-cluster-1, my-cluster-2

Example: my-snapshot-20131010

$sel:severity:CreateEventSubscription', createEventSubscription_severity - Specifies the Amazon Redshift event severity to be published by the event notification subscription.

Values: ERROR, INFO

$sel:enabled:CreateEventSubscription', createEventSubscription_enabled - A boolean value; set to true to activate the subscription, and set to false to create the subscription but not activate it.

$sel:eventCategories:CreateEventSubscription', createEventSubscription_eventCategories - Specifies the Amazon Redshift event categories to be published by the event notification subscription.

Values: configuration, management, monitoring, security, pending

$sel:tags:CreateEventSubscription', createEventSubscription_tags - A list of tag instances.

$sel:sourceType:CreateEventSubscription', createEventSubscription_sourceType - The type of source that will be generating the events. For example, if you want to be notified of events generated by a cluster, you would set this parameter to cluster. If this value is not specified, events are returned for all Amazon Redshift objects in your Amazon Web Services account. You must specify a source type in order to specify source IDs.

Valid values: cluster, cluster-parameter-group, cluster-security-group, cluster-snapshot, and scheduled-action.

$sel:subscriptionName:CreateEventSubscription', createEventSubscription_subscriptionName - The name of the event subscription to be created.

Constraints:

Cannot be null, empty, or blank.

Must contain from 1 to 255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

$sel:snsTopicArn:CreateEventSubscription', createEventSubscription_snsTopicArn - The Amazon Resource Name (ARN) of the Amazon SNS topic used to transmit the event notifications. The ARN is created by Amazon SNS when you create a topic and subscribe to it.

Request Lenses

createEventSubscription_sourceIds :: Lens' CreateEventSubscription (Maybe [Text])

A list of one or more identifiers of Amazon Redshift source objects. All of the objects must be of the same type as was specified in the source type parameter. The event subscription will return only events generated by the specified objects. If not specified, then events are returned for all objects within the source type specified.

Example: my-cluster-1, my-cluster-2

Example: my-snapshot-20131010

createEventSubscription_severity :: Lens' CreateEventSubscription (Maybe Text)

Specifies the Amazon Redshift event severity to be published by the event notification subscription.

Values: ERROR, INFO

createEventSubscription_enabled :: Lens' CreateEventSubscription (Maybe Bool)

A boolean value; set to true to activate the subscription, and set to false to create the subscription but not activate it.

createEventSubscription_eventCategories :: Lens' CreateEventSubscription (Maybe [Text])

Specifies the Amazon Redshift event categories to be published by the event notification subscription.

Values: configuration, management, monitoring, security, pending

createEventSubscription_tags :: Lens' CreateEventSubscription (Maybe [Tag])

A list of tag instances.

createEventSubscription_sourceType :: Lens' CreateEventSubscription (Maybe Text)

The type of source that will be generating the events. For example, if you want to be notified of events generated by a cluster, you would set this parameter to cluster. If this value is not specified, events are returned for all Amazon Redshift objects in your Amazon Web Services account. You must specify a source type in order to specify source IDs.

Valid values: cluster, cluster-parameter-group, cluster-security-group, cluster-snapshot, and scheduled-action.

createEventSubscription_subscriptionName :: Lens' CreateEventSubscription Text

The name of the event subscription to be created.

Constraints:

Cannot be null, empty, or blank.

Must contain from 1 to 255 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

createEventSubscription_snsTopicArn :: Lens' CreateEventSubscription Text

The Amazon Resource Name (ARN) of the Amazon SNS topic used to transmit the event notifications. The ARN is created by Amazon SNS when you create a topic and subscribe to it.

Destructuring the Response

data CreateEventSubscriptionResponse

See: newCreateEventSubscriptionResponse smart constructor.

Constructors

CreateEventSubscriptionResponse'

Fields

eventSubscription :: Maybe EventSubscription

httpStatus :: Int

The response's http status code.

newCreateEventSubscriptionResponse

Arguments

:: Int
$sel:httpStatus:CreateEventSubscriptionResponse'

-> CreateEventSubscriptionResponse

Create a value of CreateEventSubscriptionResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventSubscription:CreateEventSubscriptionResponse', createEventSubscriptionResponse_eventSubscription - Undocumented member.

$sel:httpStatus:CreateEventSubscriptionResponse', createEventSubscriptionResponse_httpStatus - The response's http status code.

Response Lenses

createEventSubscriptionResponse_eventSubscription :: Lens' CreateEventSubscriptionResponse (Maybe EventSubscription)

Undocumented member.

createEventSubscriptionResponse_httpStatus :: Lens' CreateEventSubscriptionResponse Int

The response's http status code.