Network.AWS.Redshift.CreateHsmClientCertificate

Description

Creates an HSM client certificate that an Amazon Redshift cluster will use to connect to the client's HSM in order to store and retrieve the keys used to encrypt the cluster databases.

The command returns a public key, which you must store in the HSM. In addition to creating the HSM certificate, you must create an Amazon Redshift HSM configuration that provides a cluster the information needed to store and use encryption keys in the HSM. For more information, go to Hardware Security Modules in the Amazon Redshift Cluster Management Guide.

Creating a Request

data CreateHsmClientCertificate

See: newCreateHsmClientCertificate smart constructor.

Constructors

CreateHsmClientCertificate'

Fields

tags :: Maybe [Tag]

A list of tag instances.

hsmClientCertificateIdentifier :: Text

The identifier to be assigned to the new HSM client certificate that the cluster will use to connect to the HSM to use the database encryption keys.

newCreateHsmClientCertificate

Arguments

:: Text
$sel:hsmClientCertificateIdentifier:CreateHsmClientCertificate'

-> CreateHsmClientCertificate

Create a value of CreateHsmClientCertificate with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateHsmClientCertificate', createHsmClientCertificate_tags - A list of tag instances.

$sel:hsmClientCertificateIdentifier:CreateHsmClientCertificate', createHsmClientCertificate_hsmClientCertificateIdentifier - The identifier to be assigned to the new HSM client certificate that the cluster will use to connect to the HSM to use the database encryption keys.

Request Lenses

createHsmClientCertificate_tags :: Lens' CreateHsmClientCertificate (Maybe [Tag])

A list of tag instances.

createHsmClientCertificate_hsmClientCertificateIdentifier :: Lens' CreateHsmClientCertificate Text

The identifier to be assigned to the new HSM client certificate that the cluster will use to connect to the HSM to use the database encryption keys.

Destructuring the Response

data CreateHsmClientCertificateResponse

See: newCreateHsmClientCertificateResponse smart constructor.

Constructors

CreateHsmClientCertificateResponse'

Fields

hsmClientCertificate :: Maybe HsmClientCertificate

httpStatus :: Int

The response's http status code.

newCreateHsmClientCertificateResponse

Arguments

:: Int
$sel:httpStatus:CreateHsmClientCertificateResponse'

-> CreateHsmClientCertificateResponse

Create a value of CreateHsmClientCertificateResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hsmClientCertificate:CreateHsmClientCertificateResponse', createHsmClientCertificateResponse_hsmClientCertificate - Undocumented member.

$sel:httpStatus:CreateHsmClientCertificateResponse', createHsmClientCertificateResponse_httpStatus - The response's http status code.

Response Lenses

createHsmClientCertificateResponse_hsmClientCertificate :: Lens' CreateHsmClientCertificateResponse (Maybe HsmClientCertificate)

Undocumented member.

createHsmClientCertificateResponse_httpStatus :: Lens' CreateHsmClientCertificateResponse Int

The response's http status code.