Network.AWS.Redshift.CreateHsmConfiguration

Description

Creates an HSM configuration that contains the information required by an Amazon Redshift cluster to store and use database encryption keys in a Hardware Security Module (HSM). After creating the HSM configuration, you can specify it as a parameter when creating a cluster. The cluster will then store its encryption keys in the HSM.

In addition to creating an HSM configuration, you must also create an HSM client certificate. For more information, go to Hardware Security Modules in the Amazon Redshift Cluster Management Guide.

Creating a Request

data CreateHsmConfiguration

See: newCreateHsmConfiguration smart constructor.

Constructors

CreateHsmConfiguration'

Fields

tags :: Maybe [Tag]

A list of tag instances.

hsmConfigurationIdentifier :: Text

The identifier to be assigned to the new Amazon Redshift HSM configuration.

description :: Text

A text description of the HSM configuration to be created.

hsmIpAddress :: Text

The IP address that the Amazon Redshift cluster must use to access the HSM.

hsmPartitionName :: Text

The name of the partition in the HSM where the Amazon Redshift clusters will store their database encryption keys.

hsmPartitionPassword :: Text

The password required to access the HSM partition.

hsmServerPublicCertificate :: Text

The HSMs public certificate file. When using Cloud HSM, the file name is server.pem.

newCreateHsmConfiguration

Arguments

:: Text
$sel:hsmConfigurationIdentifier:CreateHsmConfiguration'

-> Text
$sel:description:CreateHsmConfiguration'

-> Text
$sel:hsmIpAddress:CreateHsmConfiguration'

-> Text
$sel:hsmPartitionName:CreateHsmConfiguration'

-> Text
$sel:hsmPartitionPassword:CreateHsmConfiguration'

-> Text
$sel:hsmServerPublicCertificate:CreateHsmConfiguration'

-> CreateHsmConfiguration

Create a value of CreateHsmConfiguration with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tags:CreateHsmConfiguration', createHsmConfiguration_tags - A list of tag instances.

$sel:hsmConfigurationIdentifier:CreateHsmConfiguration', createHsmConfiguration_hsmConfigurationIdentifier - The identifier to be assigned to the new Amazon Redshift HSM configuration.

$sel:description:CreateHsmConfiguration', createHsmConfiguration_description - A text description of the HSM configuration to be created.

$sel:hsmIpAddress:CreateHsmConfiguration', createHsmConfiguration_hsmIpAddress - The IP address that the Amazon Redshift cluster must use to access the HSM.

$sel:hsmPartitionName:CreateHsmConfiguration', createHsmConfiguration_hsmPartitionName - The name of the partition in the HSM where the Amazon Redshift clusters will store their database encryption keys.

$sel:hsmPartitionPassword:CreateHsmConfiguration', createHsmConfiguration_hsmPartitionPassword - The password required to access the HSM partition.

$sel:hsmServerPublicCertificate:CreateHsmConfiguration', createHsmConfiguration_hsmServerPublicCertificate - The HSMs public certificate file. When using Cloud HSM, the file name is server.pem.

Request Lenses

createHsmConfiguration_tags :: Lens' CreateHsmConfiguration (Maybe [Tag])

A list of tag instances.

createHsmConfiguration_hsmConfigurationIdentifier :: Lens' CreateHsmConfiguration Text

The identifier to be assigned to the new Amazon Redshift HSM configuration.

createHsmConfiguration_description :: Lens' CreateHsmConfiguration Text

A text description of the HSM configuration to be created.

createHsmConfiguration_hsmIpAddress :: Lens' CreateHsmConfiguration Text

The IP address that the Amazon Redshift cluster must use to access the HSM.

createHsmConfiguration_hsmPartitionName :: Lens' CreateHsmConfiguration Text

The name of the partition in the HSM where the Amazon Redshift clusters will store their database encryption keys.

createHsmConfiguration_hsmPartitionPassword :: Lens' CreateHsmConfiguration Text

The password required to access the HSM partition.

createHsmConfiguration_hsmServerPublicCertificate :: Lens' CreateHsmConfiguration Text

The HSMs public certificate file. When using Cloud HSM, the file name is server.pem.

Destructuring the Response

data CreateHsmConfigurationResponse

See: newCreateHsmConfigurationResponse smart constructor.

Constructors

CreateHsmConfigurationResponse'

Fields

hsmConfiguration :: Maybe HsmConfiguration

httpStatus :: Int

The response's http status code.

newCreateHsmConfigurationResponse

Arguments

:: Int
$sel:httpStatus:CreateHsmConfigurationResponse'

-> CreateHsmConfigurationResponse

Create a value of CreateHsmConfigurationResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:hsmConfiguration:CreateHsmConfigurationResponse', createHsmConfigurationResponse_hsmConfiguration - Undocumented member.

$sel:httpStatus:CreateHsmConfigurationResponse', createHsmConfigurationResponse_httpStatus - The response's http status code.

Response Lenses

createHsmConfigurationResponse_hsmConfiguration :: Lens' CreateHsmConfigurationResponse (Maybe HsmConfiguration)

Undocumented member.

createHsmConfigurationResponse_httpStatus :: Lens' CreateHsmConfigurationResponse Int

The response's http status code.