Network.AWS.Redshift.CreateUsageLimit

Description

Creates a usage limit for a specified Amazon Redshift feature on a cluster. The usage limit is identified by the returned usage limit identifier.

Creating a Request

data CreateUsageLimit

See: newCreateUsageLimit smart constructor.

Constructors

CreateUsageLimit'

Fields

breachAction :: Maybe UsageLimitBreachAction

The action that Amazon Redshift takes when the limit is reached. The default is log. For more information about this parameter, see UsageLimit.

tags :: Maybe [Tag]

A list of tag instances.

period :: Maybe UsageLimitPeriod

The time period that the amount applies to. A weekly period begins on Sunday. The default is monthly.

clusterIdentifier :: Text

The identifier of the cluster that you want to limit usage.

featureType :: UsageLimitFeatureType

The Amazon Redshift feature that you want to limit.

limitType :: UsageLimitLimitType

The type of limit. Depending on the feature type, this can be based on a time duration or data size. If FeatureType is spectrum, then LimitType must be data-scanned. If FeatureType is concurrency-scaling, then LimitType must be time.

amount :: Integer

The limit amount. If time-based, this amount is in minutes. If data-based, this amount is in terabytes (TB). The value must be a positive number.

newCreateUsageLimit

Arguments

:: Text
$sel:clusterIdentifier:CreateUsageLimit'

-> UsageLimitFeatureType
$sel:featureType:CreateUsageLimit'

-> UsageLimitLimitType
$sel:limitType:CreateUsageLimit'

-> Integer
$sel:amount:CreateUsageLimit'

-> CreateUsageLimit

Create a value of CreateUsageLimit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:breachAction:CreateUsageLimit', createUsageLimit_breachAction - The action that Amazon Redshift takes when the limit is reached. The default is log. For more information about this parameter, see UsageLimit.

$sel:tags:CreateUsageLimit', createUsageLimit_tags - A list of tag instances.

$sel:period:CreateUsageLimit', createUsageLimit_period - The time period that the amount applies to. A weekly period begins on Sunday. The default is monthly.

$sel:clusterIdentifier:CreateUsageLimit', createUsageLimit_clusterIdentifier - The identifier of the cluster that you want to limit usage.

$sel:featureType:CreateUsageLimit', createUsageLimit_featureType - The Amazon Redshift feature that you want to limit.

$sel:limitType:CreateUsageLimit', createUsageLimit_limitType - The type of limit. Depending on the feature type, this can be based on a time duration or data size. If FeatureType is spectrum, then LimitType must be data-scanned. If FeatureType is concurrency-scaling, then LimitType must be time.

$sel:amount:CreateUsageLimit', createUsageLimit_amount - The limit amount. If time-based, this amount is in minutes. If data-based, this amount is in terabytes (TB). The value must be a positive number.

Request Lenses

createUsageLimit_breachAction :: Lens' CreateUsageLimit (Maybe UsageLimitBreachAction)

The action that Amazon Redshift takes when the limit is reached. The default is log. For more information about this parameter, see UsageLimit.

createUsageLimit_tags :: Lens' CreateUsageLimit (Maybe [Tag])

A list of tag instances.

createUsageLimit_period :: Lens' CreateUsageLimit (Maybe UsageLimitPeriod)

The time period that the amount applies to. A weekly period begins on Sunday. The default is monthly.

createUsageLimit_clusterIdentifier :: Lens' CreateUsageLimit Text

The identifier of the cluster that you want to limit usage.

createUsageLimit_featureType :: Lens' CreateUsageLimit UsageLimitFeatureType

The Amazon Redshift feature that you want to limit.

createUsageLimit_limitType :: Lens' CreateUsageLimit UsageLimitLimitType

The type of limit. Depending on the feature type, this can be based on a time duration or data size. If FeatureType is spectrum, then LimitType must be data-scanned. If FeatureType is concurrency-scaling, then LimitType must be time.

createUsageLimit_amount :: Lens' CreateUsageLimit Integer

The limit amount. If time-based, this amount is in minutes. If data-based, this amount is in terabytes (TB). The value must be a positive number.

Destructuring the Response

data UsageLimit

Describes a usage limit object for a cluster.

See: newUsageLimit smart constructor.

Constructors

UsageLimit'

Fields

amount :: Maybe Integer

The limit amount. If time-based, this amount is in minutes. If data-based, this amount is in terabytes (TB).

featureType :: Maybe UsageLimitFeatureType

The Amazon Redshift feature to which the limit applies.

breachAction :: Maybe UsageLimitBreachAction

The action that Amazon Redshift takes when the limit is reached. Possible values are:

log - To log an event in a system table. The default is log.

emit-metric - To emit CloudWatch metrics.

disable - To disable the feature until the next usage period begins.

limitType :: Maybe UsageLimitLimitType

The type of limit. Depending on the feature type, this can be based on a time duration or data size.

tags :: Maybe [Tag]

A list of tag instances.

clusterIdentifier :: Maybe Text

The identifier of the cluster with a usage limit.

period :: Maybe UsageLimitPeriod

The time period that the amount applies to. A weekly period begins on Sunday. The default is monthly.

usageLimitId :: Maybe Text

The identifier of the usage limit.

newUsageLimit :: UsageLimit

Create a value of UsageLimit with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:amount:UsageLimit', usageLimit_amount - The limit amount. If time-based, this amount is in minutes. If data-based, this amount is in terabytes (TB).

$sel:featureType:UsageLimit', usageLimit_featureType - The Amazon Redshift feature to which the limit applies.

$sel:breachAction:UsageLimit', usageLimit_breachAction - The action that Amazon Redshift takes when the limit is reached. Possible values are:

log - To log an event in a system table. The default is log.

emit-metric - To emit CloudWatch metrics.

disable - To disable the feature until the next usage period begins.

$sel:limitType:UsageLimit', usageLimit_limitType - The type of limit. Depending on the feature type, this can be based on a time duration or data size.

$sel:tags:UsageLimit', usageLimit_tags - A list of tag instances.

$sel:clusterIdentifier:UsageLimit', usageLimit_clusterIdentifier - The identifier of the cluster with a usage limit.

$sel:period:UsageLimit', usageLimit_period - The time period that the amount applies to. A weekly period begins on Sunday. The default is monthly.

$sel:usageLimitId:UsageLimit', usageLimit_usageLimitId - The identifier of the usage limit.

Response Lenses

usageLimit_amount :: Lens' UsageLimit (Maybe Integer)

The limit amount. If time-based, this amount is in minutes. If data-based, this amount is in terabytes (TB).

usageLimit_featureType :: Lens' UsageLimit (Maybe UsageLimitFeatureType)

The Amazon Redshift feature to which the limit applies.

usageLimit_breachAction :: Lens' UsageLimit (Maybe UsageLimitBreachAction)

The action that Amazon Redshift takes when the limit is reached. Possible values are:

log - To log an event in a system table. The default is log.

emit-metric - To emit CloudWatch metrics.

disable - To disable the feature until the next usage period begins.

usageLimit_limitType :: Lens' UsageLimit (Maybe UsageLimitLimitType)

The type of limit. Depending on the feature type, this can be based on a time duration or data size.

usageLimit_tags :: Lens' UsageLimit (Maybe [Tag])

A list of tag instances.

usageLimit_clusterIdentifier :: Lens' UsageLimit (Maybe Text)

The identifier of the cluster with a usage limit.

usageLimit_period :: Lens' UsageLimit (Maybe UsageLimitPeriod)

The time period that the amount applies to. A weekly period begins on Sunday. The default is monthly.

usageLimit_usageLimitId :: Lens' UsageLimit (Maybe Text)

The identifier of the usage limit.