Network.AWS.Redshift.DeleteCluster

Description

Deletes a previously provisioned cluster without its final snapshot being created. A successful response from the web service indicates that the request was received correctly. Use DescribeClusters to monitor the status of the deletion. The delete operation cannot be canceled or reverted once submitted. For more information about managing clusters, go to Amazon Redshift Clusters in the Amazon Redshift Cluster Management Guide.

If you want to shut down the cluster and retain it for future use, set SkipFinalClusterSnapshot to false and specify a name for FinalClusterSnapshotIdentifier. You can later restore this snapshot to resume using the cluster. If a final cluster snapshot is requested, the status of the cluster will be "final-snapshot" while the snapshot is being taken, then it's "deleting" once Amazon Redshift begins deleting the cluster.

For more information about managing clusters, go to Amazon Redshift Clusters in the Amazon Redshift Cluster Management Guide.

Creating a Request

data DeleteCluster

See: newDeleteCluster smart constructor.

Constructors

DeleteCluster'

Fields

skipFinalClusterSnapshot :: Maybe Bool

Determines whether a final snapshot of the cluster is created before Amazon Redshift deletes the cluster. If true, a final cluster snapshot is not created. If false, a final cluster snapshot is created before the cluster is deleted.

The FinalClusterSnapshotIdentifier parameter must be specified if SkipFinalClusterSnapshot is false.

Default: false

finalClusterSnapshotRetentionPeriod :: Maybe Int

The number of days that a manual snapshot is retained. If the value is -1, the manual snapshot is retained indefinitely.

The value must be either -1 or an integer between 1 and 3,653.

The default value is -1.

finalClusterSnapshotIdentifier :: Maybe Text

The identifier of the final snapshot that is to be created immediately before deleting the cluster. If this parameter is provided, SkipFinalClusterSnapshot must be false.

Constraints:

Must be 1 to 255 alphanumeric characters.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

clusterIdentifier :: Text

The identifier of the cluster to be deleted.

Constraints:

Must contain lowercase characters.

Must contain from 1 to 63 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

newDeleteCluster

Arguments

:: Text
$sel:clusterIdentifier:DeleteCluster'

-> DeleteCluster

Create a value of DeleteCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:skipFinalClusterSnapshot:DeleteCluster', deleteCluster_skipFinalClusterSnapshot - Determines whether a final snapshot of the cluster is created before Amazon Redshift deletes the cluster. If true, a final cluster snapshot is not created. If false, a final cluster snapshot is created before the cluster is deleted.

The FinalClusterSnapshotIdentifier parameter must be specified if SkipFinalClusterSnapshot is false.

Default: false

$sel:finalClusterSnapshotRetentionPeriod:DeleteCluster', deleteCluster_finalClusterSnapshotRetentionPeriod - The number of days that a manual snapshot is retained. If the value is -1, the manual snapshot is retained indefinitely.

The value must be either -1 or an integer between 1 and 3,653.

The default value is -1.

$sel:finalClusterSnapshotIdentifier:DeleteCluster', deleteCluster_finalClusterSnapshotIdentifier - The identifier of the final snapshot that is to be created immediately before deleting the cluster. If this parameter is provided, SkipFinalClusterSnapshot must be false.

Constraints:

Must be 1 to 255 alphanumeric characters.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

$sel:clusterIdentifier:DeleteCluster', deleteCluster_clusterIdentifier - The identifier of the cluster to be deleted.

Constraints:

Must contain lowercase characters.

Must contain from 1 to 63 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Request Lenses

deleteCluster_skipFinalClusterSnapshot :: Lens' DeleteCluster (Maybe Bool)

Determines whether a final snapshot of the cluster is created before Amazon Redshift deletes the cluster. If true, a final cluster snapshot is not created. If false, a final cluster snapshot is created before the cluster is deleted.

The FinalClusterSnapshotIdentifier parameter must be specified if SkipFinalClusterSnapshot is false.

Default: false

deleteCluster_finalClusterSnapshotRetentionPeriod :: Lens' DeleteCluster (Maybe Int)

The number of days that a manual snapshot is retained. If the value is -1, the manual snapshot is retained indefinitely.

The value must be either -1 or an integer between 1 and 3,653.

The default value is -1.

deleteCluster_finalClusterSnapshotIdentifier :: Lens' DeleteCluster (Maybe Text)

The identifier of the final snapshot that is to be created immediately before deleting the cluster. If this parameter is provided, SkipFinalClusterSnapshot must be false.

Constraints:

Must be 1 to 255 alphanumeric characters.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

deleteCluster_clusterIdentifier :: Lens' DeleteCluster Text

The identifier of the cluster to be deleted.

Constraints:

Must contain lowercase characters.

Must contain from 1 to 63 alphanumeric characters or hyphens.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Destructuring the Response

data DeleteClusterResponse

See: newDeleteClusterResponse smart constructor.

Constructors

DeleteClusterResponse'

Fields

cluster :: Maybe Cluster

httpStatus :: Int

The response's http status code.

newDeleteClusterResponse

Arguments

:: Int
$sel:httpStatus:DeleteClusterResponse'

-> DeleteClusterResponse

Create a value of DeleteClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:DeleteClusterResponse', deleteClusterResponse_cluster - Undocumented member.

$sel:httpStatus:DeleteClusterResponse', deleteClusterResponse_httpStatus - The response's http status code.

Response Lenses

deleteClusterResponse_cluster :: Lens' DeleteClusterResponse (Maybe Cluster)

Undocumented member.

deleteClusterResponse_httpStatus :: Lens' DeleteClusterResponse Int

The response's http status code.