Network.AWS.Redshift.DeleteClusterSecurityGroup

Description

Deletes an Amazon Redshift security group.

You cannot delete a security group that is associated with any clusters. You cannot delete the default security group.

For information about managing security groups, go to Amazon Redshift Cluster Security Groups in the Amazon Redshift Cluster Management Guide.

Creating a Request

data DeleteClusterSecurityGroup

See: newDeleteClusterSecurityGroup smart constructor.

Constructors

DeleteClusterSecurityGroup'

Fields

clusterSecurityGroupName :: Text

The name of the cluster security group to be deleted.

newDeleteClusterSecurityGroup

Arguments

:: Text
$sel:clusterSecurityGroupName:DeleteClusterSecurityGroup'

-> DeleteClusterSecurityGroup

Create a value of DeleteClusterSecurityGroup with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterSecurityGroupName:DeleteClusterSecurityGroup', deleteClusterSecurityGroup_clusterSecurityGroupName - The name of the cluster security group to be deleted.

Request Lenses

deleteClusterSecurityGroup_clusterSecurityGroupName :: Lens' DeleteClusterSecurityGroup Text

The name of the cluster security group to be deleted.

Destructuring the Response

data DeleteClusterSecurityGroupResponse

See: newDeleteClusterSecurityGroupResponse smart constructor.

Constructors

DeleteClusterSecurityGroupResponse'

newDeleteClusterSecurityGroupResponse :: DeleteClusterSecurityGroupResponse

Create a value of DeleteClusterSecurityGroupResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.