Network.AWS.Redshift.DescribeClusterParameters

Description

Returns a detailed list of parameters contained within the specified Amazon Redshift parameter group. For each parameter the response includes information such as parameter name, description, data type, value, whether the parameter value is modifiable, and so on.

You can specify source filter to retrieve parameters of only specific type. For example, to retrieve parameters that were modified by a user action such as from ModifyClusterParameterGroup, you can specify source equal to user.

For more information about parameters and parameter groups, go to Amazon Redshift Parameter Groups in the Amazon Redshift Cluster Management Guide.

This operation returns paginated results.

Creating a Request

data DescribeClusterParameters

See: newDescribeClusterParameters smart constructor.

Constructors

DescribeClusterParameters'

Fields

source :: Maybe Text

The parameter types to return. Specify user to show parameters that are different form the default. Similarly, specify engine-default to show parameters that are the same as the default parameter group.

Default: All parameter types returned.

Valid Values: user | engine-default

maxRecords :: Maybe Int

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

marker :: Maybe Text

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterParameters request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

parameterGroupName :: Text

The name of a cluster parameter group for which to return details.

newDescribeClusterParameters

Arguments

:: Text
$sel:parameterGroupName:DescribeClusterParameters'

-> DescribeClusterParameters

Create a value of DescribeClusterParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:source:DescribeClusterParameters', describeClusterParameters_source - The parameter types to return. Specify user to show parameters that are different form the default. Similarly, specify engine-default to show parameters that are the same as the default parameter group.

Default: All parameter types returned.

Valid Values: user | engine-default

$sel:maxRecords:DescribeClusterParameters', describeClusterParameters_maxRecords - The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

$sel:marker:DescribeClusterParameters', describeClusterParameters_marker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterParameters request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

$sel:parameterGroupName:DescribeClusterParameters', describeClusterParameters_parameterGroupName - The name of a cluster parameter group for which to return details.

Request Lenses

describeClusterParameters_source :: Lens' DescribeClusterParameters (Maybe Text)

The parameter types to return. Specify user to show parameters that are different form the default. Similarly, specify engine-default to show parameters that are the same as the default parameter group.

Default: All parameter types returned.

Valid Values: user | engine-default

describeClusterParameters_maxRecords :: Lens' DescribeClusterParameters (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

describeClusterParameters_marker :: Lens' DescribeClusterParameters (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterParameters request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

describeClusterParameters_parameterGroupName :: Lens' DescribeClusterParameters Text

The name of a cluster parameter group for which to return details.

Destructuring the Response

data DescribeClusterParametersResponse

Contains the output from the DescribeClusterParameters action.

See: newDescribeClusterParametersResponse smart constructor.

Constructors

DescribeClusterParametersResponse'

Fields

parameters :: Maybe [Parameter]

A list of Parameter instances. Each instance lists the parameters of one cluster parameter group.

marker :: Maybe Text

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

httpStatus :: Int

The response's http status code.

newDescribeClusterParametersResponse

Arguments

:: Int
$sel:httpStatus:DescribeClusterParametersResponse'

-> DescribeClusterParametersResponse

Create a value of DescribeClusterParametersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:parameters:DescribeClusterParametersResponse', describeClusterParametersResponse_parameters - A list of Parameter instances. Each instance lists the parameters of one cluster parameter group.

$sel:marker:DescribeClusterParameters', describeClusterParametersResponse_marker - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

$sel:httpStatus:DescribeClusterParametersResponse', describeClusterParametersResponse_httpStatus - The response's http status code.

Response Lenses

describeClusterParametersResponse_parameters :: Lens' DescribeClusterParametersResponse (Maybe [Parameter])

A list of Parameter instances. Each instance lists the parameters of one cluster parameter group.

describeClusterParametersResponse_marker :: Lens' DescribeClusterParametersResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

describeClusterParametersResponse_httpStatus :: Lens' DescribeClusterParametersResponse Int

The response's http status code.