Network.AWS.Redshift.DescribeClusterSecurityGroups

Description

Returns information about Amazon Redshift security groups. If the name of a security group is specified, the response will contain only information about only that security group.

For information about managing security groups, go to Amazon Redshift Cluster Security Groups in the Amazon Redshift Cluster Management Guide.

If you specify both tag keys and tag values in the same request, Amazon Redshift returns all security groups that match any combination of the specified keys and values. For example, if you have owner and environment for tag keys, and admin and test for tag values, all security groups that have any combination of those values are returned.

If both tag keys and values are omitted from the request, security groups are returned regardless of whether they have tag keys or values associated with them.

This operation returns paginated results.

Creating a Request

data DescribeClusterSecurityGroups

See: newDescribeClusterSecurityGroups smart constructor.

Constructors

DescribeClusterSecurityGroups'

Fields

tagKeys :: Maybe [Text]

A tag key or keys for which you want to return all matching cluster security groups that are associated with the specified key or keys. For example, suppose that you have security groups that are tagged with keys called owner and environment. If you specify both of these tag keys in the request, Amazon Redshift returns a response with the security groups that have either or both of these tag keys associated with them.

clusterSecurityGroupName :: Maybe Text

The name of a cluster security group for which you are requesting details. You can specify either the Marker parameter or a ClusterSecurityGroupName parameter, but not both.

Example: securitygroup1

tagValues :: Maybe [Text]

A tag value or values for which you want to return all matching cluster security groups that are associated with the specified tag value or values. For example, suppose that you have security groups that are tagged with values called admin and test. If you specify both of these tag values in the request, Amazon Redshift returns a response with the security groups that have either or both of these tag values associated with them.

maxRecords :: Maybe Int

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

marker :: Maybe Text

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterSecurityGroups request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

Constraints: You can specify either the ClusterSecurityGroupName parameter or the Marker parameter, but not both.

newDescribeClusterSecurityGroups :: DescribeClusterSecurityGroups

Create a value of DescribeClusterSecurityGroups with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagKeys:DescribeClusterSecurityGroups', describeClusterSecurityGroups_tagKeys - A tag key or keys for which you want to return all matching cluster security groups that are associated with the specified key or keys. For example, suppose that you have security groups that are tagged with keys called owner and environment. If you specify both of these tag keys in the request, Amazon Redshift returns a response with the security groups that have either or both of these tag keys associated with them.

$sel:clusterSecurityGroupName:DescribeClusterSecurityGroups', describeClusterSecurityGroups_clusterSecurityGroupName - The name of a cluster security group for which you are requesting details. You can specify either the Marker parameter or a ClusterSecurityGroupName parameter, but not both.

Example: securitygroup1

$sel:tagValues:DescribeClusterSecurityGroups', describeClusterSecurityGroups_tagValues - A tag value or values for which you want to return all matching cluster security groups that are associated with the specified tag value or values. For example, suppose that you have security groups that are tagged with values called admin and test. If you specify both of these tag values in the request, Amazon Redshift returns a response with the security groups that have either or both of these tag values associated with them.

$sel:maxRecords:DescribeClusterSecurityGroups', describeClusterSecurityGroups_maxRecords - The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

$sel:marker:DescribeClusterSecurityGroups', describeClusterSecurityGroups_marker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterSecurityGroups request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

Constraints: You can specify either the ClusterSecurityGroupName parameter or the Marker parameter, but not both.

Request Lenses

describeClusterSecurityGroups_tagKeys :: Lens' DescribeClusterSecurityGroups (Maybe [Text])

A tag key or keys for which you want to return all matching cluster security groups that are associated with the specified key or keys. For example, suppose that you have security groups that are tagged with keys called owner and environment. If you specify both of these tag keys in the request, Amazon Redshift returns a response with the security groups that have either or both of these tag keys associated with them.

describeClusterSecurityGroups_clusterSecurityGroupName :: Lens' DescribeClusterSecurityGroups (Maybe Text)

The name of a cluster security group for which you are requesting details. You can specify either the Marker parameter or a ClusterSecurityGroupName parameter, but not both.

Example: securitygroup1

describeClusterSecurityGroups_tagValues :: Lens' DescribeClusterSecurityGroups (Maybe [Text])

A tag value or values for which you want to return all matching cluster security groups that are associated with the specified tag value or values. For example, suppose that you have security groups that are tagged with values called admin and test. If you specify both of these tag values in the request, Amazon Redshift returns a response with the security groups that have either or both of these tag values associated with them.

describeClusterSecurityGroups_maxRecords :: Lens' DescribeClusterSecurityGroups (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

describeClusterSecurityGroups_marker :: Lens' DescribeClusterSecurityGroups (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeClusterSecurityGroups request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

Constraints: You can specify either the ClusterSecurityGroupName parameter or the Marker parameter, but not both.

Destructuring the Response

data DescribeClusterSecurityGroupsResponse

See: newDescribeClusterSecurityGroupsResponse smart constructor.

Constructors

DescribeClusterSecurityGroupsResponse'

Fields

clusterSecurityGroups :: Maybe [ClusterSecurityGroup]

A list of ClusterSecurityGroup instances.

marker :: Maybe Text

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

httpStatus :: Int

The response's http status code.

newDescribeClusterSecurityGroupsResponse

Arguments

:: Int
$sel:httpStatus:DescribeClusterSecurityGroupsResponse'

-> DescribeClusterSecurityGroupsResponse

Create a value of DescribeClusterSecurityGroupsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:clusterSecurityGroups:DescribeClusterSecurityGroupsResponse', describeClusterSecurityGroupsResponse_clusterSecurityGroups - A list of ClusterSecurityGroup instances.

$sel:marker:DescribeClusterSecurityGroups', describeClusterSecurityGroupsResponse_marker - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

$sel:httpStatus:DescribeClusterSecurityGroupsResponse', describeClusterSecurityGroupsResponse_httpStatus - The response's http status code.

Response Lenses

describeClusterSecurityGroupsResponse_clusterSecurityGroups :: Lens' DescribeClusterSecurityGroupsResponse (Maybe [ClusterSecurityGroup])

A list of ClusterSecurityGroup instances.

describeClusterSecurityGroupsResponse_marker :: Lens' DescribeClusterSecurityGroupsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

describeClusterSecurityGroupsResponse_httpStatus :: Lens' DescribeClusterSecurityGroupsResponse Int

The response's http status code.