Network.AWS.Redshift.DescribeDataShares

Description

Shows the status of any inbound or outbound datashares available in the specified account.

Creating a Request

data DescribeDataShares

See: newDescribeDataShares smart constructor.

Constructors

DescribeDataShares'

Fields

dataShareArn :: Maybe Text

The identifier of the datashare to describe details of.

maxRecords :: Maybe Int

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

marker :: Maybe Text

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDataShares request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

newDescribeDataShares :: DescribeDataShares

Create a value of DescribeDataShares with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataShareArn:DescribeDataShares', describeDataShares_dataShareArn - The identifier of the datashare to describe details of.

$sel:maxRecords:DescribeDataShares', describeDataShares_maxRecords - The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

$sel:marker:DescribeDataShares', describeDataShares_marker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDataShares request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

Request Lenses

describeDataShares_dataShareArn :: Lens' DescribeDataShares (Maybe Text)

The identifier of the datashare to describe details of.

describeDataShares_maxRecords :: Lens' DescribeDataShares (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

describeDataShares_marker :: Lens' DescribeDataShares (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDataShares request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

Destructuring the Response

data DescribeDataSharesResponse

See: newDescribeDataSharesResponse smart constructor.

Constructors

DescribeDataSharesResponse'

Fields

dataShares :: Maybe [DataShare]

The results returned from describing datashares.

marker :: Maybe Text

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDataShares request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

httpStatus :: Int

The response's http status code.

newDescribeDataSharesResponse

Arguments

:: Int
$sel:httpStatus:DescribeDataSharesResponse'

-> DescribeDataSharesResponse

Create a value of DescribeDataSharesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataShares:DescribeDataSharesResponse', describeDataSharesResponse_dataShares - The results returned from describing datashares.

$sel:marker:DescribeDataShares', describeDataSharesResponse_marker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDataShares request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

$sel:httpStatus:DescribeDataSharesResponse', describeDataSharesResponse_httpStatus - The response's http status code.

Response Lenses

describeDataSharesResponse_dataShares :: Lens' DescribeDataSharesResponse (Maybe [DataShare])

The results returned from describing datashares.

describeDataSharesResponse_marker :: Lens' DescribeDataSharesResponse (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDataShares request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

describeDataSharesResponse_httpStatus :: Lens' DescribeDataSharesResponse Int

The response's http status code.