Network.AWS.Redshift.DescribeDataSharesForProducer

Description

Returns a list of datashares when the account identifier being called is a producer account identifier.

Creating a Request

data DescribeDataSharesForProducer

See: newDescribeDataSharesForProducer smart constructor.

Constructors

DescribeDataSharesForProducer'

Fields

status :: Maybe DataShareStatusForProducer

An identifier giving the status of a datashare in the producer. If this field is specified, Amazon Redshift returns the list of datashares that have the specified status.

producerArn :: Maybe Text

The Amazon Resource Name (ARN) of the producer that returns in the list of datashares.

maxRecords :: Maybe Int

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

marker :: Maybe Text

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDataSharesForProducer request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

newDescribeDataSharesForProducer :: DescribeDataSharesForProducer

Create a value of DescribeDataSharesForProducer with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:status:DescribeDataSharesForProducer', describeDataSharesForProducer_status - An identifier giving the status of a datashare in the producer. If this field is specified, Amazon Redshift returns the list of datashares that have the specified status.

$sel:producerArn:DescribeDataSharesForProducer', describeDataSharesForProducer_producerArn - The Amazon Resource Name (ARN) of the producer that returns in the list of datashares.

$sel:maxRecords:DescribeDataSharesForProducer', describeDataSharesForProducer_maxRecords - The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

$sel:marker:DescribeDataSharesForProducer', describeDataSharesForProducer_marker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDataSharesForProducer request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

Request Lenses

describeDataSharesForProducer_status :: Lens' DescribeDataSharesForProducer (Maybe DataShareStatusForProducer)

An identifier giving the status of a datashare in the producer. If this field is specified, Amazon Redshift returns the list of datashares that have the specified status.

describeDataSharesForProducer_producerArn :: Lens' DescribeDataSharesForProducer (Maybe Text)

The Amazon Resource Name (ARN) of the producer that returns in the list of datashares.

describeDataSharesForProducer_maxRecords :: Lens' DescribeDataSharesForProducer (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

describeDataSharesForProducer_marker :: Lens' DescribeDataSharesForProducer (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDataSharesForProducer request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

Destructuring the Response

data DescribeDataSharesForProducerResponse

See: newDescribeDataSharesForProducerResponse smart constructor.

Constructors

DescribeDataSharesForProducerResponse'

Fields

dataShares :: Maybe [DataShare]

Shows the results of datashares available for producers.

marker :: Maybe Text

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDataSharesForProducer request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

httpStatus :: Int

The response's http status code.

newDescribeDataSharesForProducerResponse

Arguments

:: Int
$sel:httpStatus:DescribeDataSharesForProducerResponse'

-> DescribeDataSharesForProducerResponse

Create a value of DescribeDataSharesForProducerResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dataShares:DescribeDataSharesForProducerResponse', describeDataSharesForProducerResponse_dataShares - Shows the results of datashares available for producers.

$sel:marker:DescribeDataSharesForProducer', describeDataSharesForProducerResponse_marker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDataSharesForProducer request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

$sel:httpStatus:DescribeDataSharesForProducerResponse', describeDataSharesForProducerResponse_httpStatus - The response's http status code.

Response Lenses

describeDataSharesForProducerResponse_dataShares :: Lens' DescribeDataSharesForProducerResponse (Maybe [DataShare])

Shows the results of datashares available for producers.

describeDataSharesForProducerResponse_marker :: Lens' DescribeDataSharesForProducerResponse (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDataSharesForProducer request exceed the value specified in MaxRecords, AWS returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

describeDataSharesForProducerResponse_httpStatus :: Lens' DescribeDataSharesForProducerResponse Int

The response's http status code.