Network.AWS.Redshift.DescribeDefaultClusterParameters

Description

Returns a list of parameter settings for the specified parameter group family.

For more information about parameters and parameter groups, go to Amazon Redshift Parameter Groups in the Amazon Redshift Cluster Management Guide.

This operation returns paginated results.

Creating a Request

data DescribeDefaultClusterParameters

See: newDescribeDefaultClusterParameters smart constructor.

Constructors

DescribeDefaultClusterParameters'

Fields

maxRecords :: Maybe Int

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

marker :: Maybe Text

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDefaultClusterParameters request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

parameterGroupFamily :: Text

The name of the cluster parameter group family.

newDescribeDefaultClusterParameters

Arguments

:: Text
$sel:parameterGroupFamily:DescribeDefaultClusterParameters'

-> DescribeDefaultClusterParameters

Create a value of DescribeDefaultClusterParameters with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:maxRecords:DescribeDefaultClusterParameters', describeDefaultClusterParameters_maxRecords - The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

$sel:marker:DescribeDefaultClusterParameters', describeDefaultClusterParameters_marker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDefaultClusterParameters request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

$sel:parameterGroupFamily:DescribeDefaultClusterParameters', describeDefaultClusterParameters_parameterGroupFamily - The name of the cluster parameter group family.

Request Lenses

describeDefaultClusterParameters_maxRecords :: Lens' DescribeDefaultClusterParameters (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

describeDefaultClusterParameters_marker :: Lens' DescribeDefaultClusterParameters (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeDefaultClusterParameters request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

describeDefaultClusterParameters_parameterGroupFamily :: Lens' DescribeDefaultClusterParameters Text

The name of the cluster parameter group family.

Destructuring the Response

data DescribeDefaultClusterParametersResponse

See: newDescribeDefaultClusterParametersResponse smart constructor.

Constructors

DescribeDefaultClusterParametersResponse'

Fields

httpStatus :: Int

The response's http status code.

defaultClusterParameters :: DefaultClusterParameters

newDescribeDefaultClusterParametersResponse

Arguments

:: Int
$sel:httpStatus:DescribeDefaultClusterParametersResponse'

-> DefaultClusterParameters
$sel:defaultClusterParameters:DescribeDefaultClusterParametersResponse'

-> DescribeDefaultClusterParametersResponse

Create a value of DescribeDefaultClusterParametersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:httpStatus:DescribeDefaultClusterParametersResponse', describeDefaultClusterParametersResponse_httpStatus - The response's http status code.

$sel:defaultClusterParameters:DescribeDefaultClusterParametersResponse', describeDefaultClusterParametersResponse_defaultClusterParameters - Undocumented member.

Response Lenses

describeDefaultClusterParametersResponse_httpStatus :: Lens' DescribeDefaultClusterParametersResponse Int

The response's http status code.

describeDefaultClusterParametersResponse_defaultClusterParameters :: Lens' DescribeDefaultClusterParametersResponse DefaultClusterParameters

Undocumented member.