Network.AWS.Redshift.DescribeEndpointAccess

Description

Describes a Redshift-managed VPC endpoint.

This operation returns paginated results.

Creating a Request

data DescribeEndpointAccess

See: newDescribeEndpointAccess smart constructor.

Constructors

DescribeEndpointAccess'

Fields

endpointName :: Maybe Text

The name of the endpoint to be described.

vpcId :: Maybe Text

The virtual private cloud (VPC) identifier with access to the cluster.

clusterIdentifier :: Maybe Text

The cluster identifier associated with the described endpoint.

marker :: Maybe Text

An optional pagination token provided by a previous DescribeEndpointAccess request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by the MaxRecords parameter.

maxRecords :: Maybe Int

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a Marker is included in the response so that the remaining results can be retrieved.

resourceOwner :: Maybe Text

The Amazon Web Services account ID of the owner of the cluster.

newDescribeEndpointAccess :: DescribeEndpointAccess

Create a value of DescribeEndpointAccess with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:endpointName:DescribeEndpointAccess', describeEndpointAccess_endpointName - The name of the endpoint to be described.

$sel:vpcId:DescribeEndpointAccess', describeEndpointAccess_vpcId - The virtual private cloud (VPC) identifier with access to the cluster.

$sel:clusterIdentifier:DescribeEndpointAccess', describeEndpointAccess_clusterIdentifier - The cluster identifier associated with the described endpoint.

$sel:marker:DescribeEndpointAccess', describeEndpointAccess_marker - An optional pagination token provided by a previous DescribeEndpointAccess request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by the MaxRecords parameter.

$sel:maxRecords:DescribeEndpointAccess', describeEndpointAccess_maxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a Marker is included in the response so that the remaining results can be retrieved.

$sel:resourceOwner:DescribeEndpointAccess', describeEndpointAccess_resourceOwner - The Amazon Web Services account ID of the owner of the cluster.

Request Lenses

describeEndpointAccess_endpointName :: Lens' DescribeEndpointAccess (Maybe Text)

The name of the endpoint to be described.

describeEndpointAccess_vpcId :: Lens' DescribeEndpointAccess (Maybe Text)

The virtual private cloud (VPC) identifier with access to the cluster.

describeEndpointAccess_clusterIdentifier :: Lens' DescribeEndpointAccess (Maybe Text)

The cluster identifier associated with the described endpoint.

describeEndpointAccess_marker :: Lens' DescribeEndpointAccess (Maybe Text)

An optional pagination token provided by a previous DescribeEndpointAccess request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by the MaxRecords parameter.

describeEndpointAccess_maxRecords :: Lens' DescribeEndpointAccess (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a Marker is included in the response so that the remaining results can be retrieved.

describeEndpointAccess_resourceOwner :: Lens' DescribeEndpointAccess (Maybe Text)

The Amazon Web Services account ID of the owner of the cluster.

Destructuring the Response

data DescribeEndpointAccessResponse

See: newDescribeEndpointAccessResponse smart constructor.

Constructors

DescribeEndpointAccessResponse'

Fields

marker :: Maybe Text

An optional pagination token provided by a previous DescribeEndpointAccess request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by the MaxRecords parameter.

endpointAccessList :: Maybe [EndpointAccess]

The list of endpoints with access to the cluster.

httpStatus :: Int

The response's http status code.

newDescribeEndpointAccessResponse

Arguments

:: Int
$sel:httpStatus:DescribeEndpointAccessResponse'

-> DescribeEndpointAccessResponse

Create a value of DescribeEndpointAccessResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:DescribeEndpointAccess', describeEndpointAccessResponse_marker - An optional pagination token provided by a previous DescribeEndpointAccess request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by the MaxRecords parameter.

$sel:endpointAccessList:DescribeEndpointAccessResponse', describeEndpointAccessResponse_endpointAccessList - The list of endpoints with access to the cluster.

$sel:httpStatus:DescribeEndpointAccessResponse', describeEndpointAccessResponse_httpStatus - The response's http status code.

Response Lenses

describeEndpointAccessResponse_marker :: Lens' DescribeEndpointAccessResponse (Maybe Text)

An optional pagination token provided by a previous DescribeEndpointAccess request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by the MaxRecords parameter.

describeEndpointAccessResponse_endpointAccessList :: Lens' DescribeEndpointAccessResponse (Maybe [EndpointAccess])

The list of endpoints with access to the cluster.

describeEndpointAccessResponse_httpStatus :: Lens' DescribeEndpointAccessResponse Int

The response's http status code.