Network.AWS.Redshift.DescribeEventSubscriptions

Description

Lists descriptions of all the Amazon Redshift event notification subscriptions for a customer account. If you specify a subscription name, lists the description for that subscription.

If you specify both tag keys and tag values in the same request, Amazon Redshift returns all event notification subscriptions that match any combination of the specified keys and values. For example, if you have owner and environment for tag keys, and admin and test for tag values, all subscriptions that have any combination of those values are returned.

If both tag keys and values are omitted from the request, subscriptions are returned regardless of whether they have tag keys or values associated with them.

This operation returns paginated results.

Creating a Request

data DescribeEventSubscriptions

See: newDescribeEventSubscriptions smart constructor.

Constructors

DescribeEventSubscriptions'

Fields

subscriptionName :: Maybe Text

The name of the Amazon Redshift event notification subscription to be described.

tagValues :: Maybe [Text]

A tag value or values for which you want to return all matching event notification subscriptions that are associated with the specified tag value or values. For example, suppose that you have subscriptions that are tagged with values called admin and test. If you specify both of these tag values in the request, Amazon Redshift returns a response with the subscriptions that have either or both of these tag values associated with them.

tagKeys :: Maybe [Text]

A tag key or keys for which you want to return all matching event notification subscriptions that are associated with the specified key or keys. For example, suppose that you have subscriptions that are tagged with keys called owner and environment. If you specify both of these tag keys in the request, Amazon Redshift returns a response with the subscriptions that have either or both of these tag keys associated with them.

marker :: Maybe Text

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeEventSubscriptions request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

maxRecords :: Maybe Int

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

newDescribeEventSubscriptions :: DescribeEventSubscriptions

Create a value of DescribeEventSubscriptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:subscriptionName:DescribeEventSubscriptions', describeEventSubscriptions_subscriptionName - The name of the Amazon Redshift event notification subscription to be described.

$sel:tagValues:DescribeEventSubscriptions', describeEventSubscriptions_tagValues - A tag value or values for which you want to return all matching event notification subscriptions that are associated with the specified tag value or values. For example, suppose that you have subscriptions that are tagged with values called admin and test. If you specify both of these tag values in the request, Amazon Redshift returns a response with the subscriptions that have either or both of these tag values associated with them.

$sel:tagKeys:DescribeEventSubscriptions', describeEventSubscriptions_tagKeys - A tag key or keys for which you want to return all matching event notification subscriptions that are associated with the specified key or keys. For example, suppose that you have subscriptions that are tagged with keys called owner and environment. If you specify both of these tag keys in the request, Amazon Redshift returns a response with the subscriptions that have either or both of these tag keys associated with them.

$sel:marker:DescribeEventSubscriptions', describeEventSubscriptions_marker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeEventSubscriptions request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

$sel:maxRecords:DescribeEventSubscriptions', describeEventSubscriptions_maxRecords - The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

Request Lenses

describeEventSubscriptions_subscriptionName :: Lens' DescribeEventSubscriptions (Maybe Text)

The name of the Amazon Redshift event notification subscription to be described.

describeEventSubscriptions_tagValues :: Lens' DescribeEventSubscriptions (Maybe [Text])

A tag value or values for which you want to return all matching event notification subscriptions that are associated with the specified tag value or values. For example, suppose that you have subscriptions that are tagged with values called admin and test. If you specify both of these tag values in the request, Amazon Redshift returns a response with the subscriptions that have either or both of these tag values associated with them.

describeEventSubscriptions_tagKeys :: Lens' DescribeEventSubscriptions (Maybe [Text])

A tag key or keys for which you want to return all matching event notification subscriptions that are associated with the specified key or keys. For example, suppose that you have subscriptions that are tagged with keys called owner and environment. If you specify both of these tag keys in the request, Amazon Redshift returns a response with the subscriptions that have either or both of these tag keys associated with them.

describeEventSubscriptions_marker :: Lens' DescribeEventSubscriptions (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeEventSubscriptions request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

describeEventSubscriptions_maxRecords :: Lens' DescribeEventSubscriptions (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

Destructuring the Response

data DescribeEventSubscriptionsResponse

See: newDescribeEventSubscriptionsResponse smart constructor.

Constructors

DescribeEventSubscriptionsResponse'

Fields

eventSubscriptionsList :: Maybe [EventSubscription]

A list of event subscriptions.

marker :: Maybe Text

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

httpStatus :: Int

The response's http status code.

newDescribeEventSubscriptionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeEventSubscriptionsResponse'

-> DescribeEventSubscriptionsResponse

Create a value of DescribeEventSubscriptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:eventSubscriptionsList:DescribeEventSubscriptionsResponse', describeEventSubscriptionsResponse_eventSubscriptionsList - A list of event subscriptions.

$sel:marker:DescribeEventSubscriptions', describeEventSubscriptionsResponse_marker - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

$sel:httpStatus:DescribeEventSubscriptionsResponse', describeEventSubscriptionsResponse_httpStatus - The response's http status code.

Response Lenses

describeEventSubscriptionsResponse_eventSubscriptionsList :: Lens' DescribeEventSubscriptionsResponse (Maybe [EventSubscription])

A list of event subscriptions.

describeEventSubscriptionsResponse_marker :: Lens' DescribeEventSubscriptionsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

describeEventSubscriptionsResponse_httpStatus :: Lens' DescribeEventSubscriptionsResponse Int

The response's http status code.