Network.AWS.Redshift.DescribeNodeConfigurationOptions

Description

Returns properties of possible node configurations such as node type, number of nodes, and disk usage for the specified action type.

This operation returns paginated results.

Creating a Request

data DescribeNodeConfigurationOptions

See: newDescribeNodeConfigurationOptions smart constructor.

Constructors

DescribeNodeConfigurationOptions'

Fields

snapshotIdentifier :: Maybe Text

The identifier of the snapshot to evaluate for possible node configurations.

filters :: Maybe [NodeConfigurationOptionsFilter]

A set of name, operator, and value items to filter the results.

clusterIdentifier :: Maybe Text

The identifier of the cluster to evaluate for possible node configurations.

marker :: Maybe Text

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeNodeConfigurationOptions request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

maxRecords :: Maybe Int

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 500

Constraints: minimum 100, maximum 500.

ownerAccount :: Maybe Text

The Amazon Web Services account used to create or copy the snapshot. Required if you are restoring a snapshot you do not own, optional if you own the snapshot.

actionType :: ActionType

The action type to evaluate for possible node configurations. Specify "restore-cluster" to get configuration combinations based on an existing snapshot. Specify "recommend-node-config" to get configuration recommendations based on an existing cluster or snapshot. Specify "resize-cluster" to get configuration combinations for elastic resize based on an existing cluster.

newDescribeNodeConfigurationOptions

Arguments

:: ActionType
$sel:actionType:DescribeNodeConfigurationOptions'

-> DescribeNodeConfigurationOptions

Create a value of DescribeNodeConfigurationOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshotIdentifier:DescribeNodeConfigurationOptions', describeNodeConfigurationOptions_snapshotIdentifier - The identifier of the snapshot to evaluate for possible node configurations.

$sel:filters:DescribeNodeConfigurationOptions', describeNodeConfigurationOptions_filters - A set of name, operator, and value items to filter the results.

$sel:clusterIdentifier:DescribeNodeConfigurationOptions', describeNodeConfigurationOptions_clusterIdentifier - The identifier of the cluster to evaluate for possible node configurations.

$sel:marker:DescribeNodeConfigurationOptions', describeNodeConfigurationOptions_marker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeNodeConfigurationOptions request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

$sel:maxRecords:DescribeNodeConfigurationOptions', describeNodeConfigurationOptions_maxRecords - The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 500

Constraints: minimum 100, maximum 500.

$sel:ownerAccount:DescribeNodeConfigurationOptions', describeNodeConfigurationOptions_ownerAccount - The Amazon Web Services account used to create or copy the snapshot. Required if you are restoring a snapshot you do not own, optional if you own the snapshot.

$sel:actionType:DescribeNodeConfigurationOptions', describeNodeConfigurationOptions_actionType - The action type to evaluate for possible node configurations. Specify "restore-cluster" to get configuration combinations based on an existing snapshot. Specify "recommend-node-config" to get configuration recommendations based on an existing cluster or snapshot. Specify "resize-cluster" to get configuration combinations for elastic resize based on an existing cluster.

Request Lenses

describeNodeConfigurationOptions_snapshotIdentifier :: Lens' DescribeNodeConfigurationOptions (Maybe Text)

The identifier of the snapshot to evaluate for possible node configurations.

describeNodeConfigurationOptions_filters :: Lens' DescribeNodeConfigurationOptions (Maybe [NodeConfigurationOptionsFilter])

A set of name, operator, and value items to filter the results.

describeNodeConfigurationOptions_clusterIdentifier :: Lens' DescribeNodeConfigurationOptions (Maybe Text)

The identifier of the cluster to evaluate for possible node configurations.

describeNodeConfigurationOptions_marker :: Lens' DescribeNodeConfigurationOptions (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeNodeConfigurationOptions request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

describeNodeConfigurationOptions_maxRecords :: Lens' DescribeNodeConfigurationOptions (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 500

Constraints: minimum 100, maximum 500.

describeNodeConfigurationOptions_ownerAccount :: Lens' DescribeNodeConfigurationOptions (Maybe Text)

The Amazon Web Services account used to create or copy the snapshot. Required if you are restoring a snapshot you do not own, optional if you own the snapshot.

describeNodeConfigurationOptions_actionType :: Lens' DescribeNodeConfigurationOptions ActionType

The action type to evaluate for possible node configurations. Specify "restore-cluster" to get configuration combinations based on an existing snapshot. Specify "recommend-node-config" to get configuration recommendations based on an existing cluster or snapshot. Specify "resize-cluster" to get configuration combinations for elastic resize based on an existing cluster.

Destructuring the Response

data DescribeNodeConfigurationOptionsResponse

See: newDescribeNodeConfigurationOptionsResponse smart constructor.

Constructors

DescribeNodeConfigurationOptionsResponse'

Fields

nodeConfigurationOptionList :: Maybe [NodeConfigurationOption]

A list of valid node configurations.

marker :: Maybe Text

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

httpStatus :: Int

The response's http status code.

newDescribeNodeConfigurationOptionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeNodeConfigurationOptionsResponse'

-> DescribeNodeConfigurationOptionsResponse

Create a value of DescribeNodeConfigurationOptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:nodeConfigurationOptionList:DescribeNodeConfigurationOptionsResponse', describeNodeConfigurationOptionsResponse_nodeConfigurationOptionList - A list of valid node configurations.

$sel:marker:DescribeNodeConfigurationOptions', describeNodeConfigurationOptionsResponse_marker - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

$sel:httpStatus:DescribeNodeConfigurationOptionsResponse', describeNodeConfigurationOptionsResponse_httpStatus - The response's http status code.

Response Lenses

describeNodeConfigurationOptionsResponse_nodeConfigurationOptionList :: Lens' DescribeNodeConfigurationOptionsResponse (Maybe [NodeConfigurationOption])

A list of valid node configurations.

describeNodeConfigurationOptionsResponse_marker :: Lens' DescribeNodeConfigurationOptionsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

describeNodeConfigurationOptionsResponse_httpStatus :: Lens' DescribeNodeConfigurationOptionsResponse Int

The response's http status code.