Network.AWS.Redshift.DescribeOrderableClusterOptions

Description

Returns a list of orderable cluster options. Before you create a new cluster you can use this operation to find what options are available, such as the EC2 Availability Zones (AZ) in the specific Amazon Web Services Region that you can specify, and the node types you can request. The node types differ by available storage, memory, CPU and price. With the cost involved you might want to obtain a list of cluster options in the specific region and specify values when creating a cluster. For more information about managing clusters, go to Amazon Redshift Clusters in the Amazon Redshift Cluster Management Guide.

This operation returns paginated results.

Creating a Request

data DescribeOrderableClusterOptions

See: newDescribeOrderableClusterOptions smart constructor.

Constructors

DescribeOrderableClusterOptions'

Fields

marker :: Maybe Text

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeOrderableClusterOptions request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

maxRecords :: Maybe Int

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

clusterVersion :: Maybe Text

The version filter value. Specify this parameter to show only the available offerings matching the specified version.

Default: All versions.

Constraints: Must be one of the version returned from DescribeClusterVersions.

nodeType :: Maybe Text

The node type filter value. Specify this parameter to show only the available offerings matching the specified node type.

newDescribeOrderableClusterOptions :: DescribeOrderableClusterOptions

Create a value of DescribeOrderableClusterOptions with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:DescribeOrderableClusterOptions', describeOrderableClusterOptions_marker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeOrderableClusterOptions request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

$sel:maxRecords:DescribeOrderableClusterOptions', describeOrderableClusterOptions_maxRecords - The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

$sel:clusterVersion:DescribeOrderableClusterOptions', describeOrderableClusterOptions_clusterVersion - The version filter value. Specify this parameter to show only the available offerings matching the specified version.

Default: All versions.

Constraints: Must be one of the version returned from DescribeClusterVersions.

$sel:nodeType:DescribeOrderableClusterOptions', describeOrderableClusterOptions_nodeType - The node type filter value. Specify this parameter to show only the available offerings matching the specified node type.

Request Lenses

describeOrderableClusterOptions_marker :: Lens' DescribeOrderableClusterOptions (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeOrderableClusterOptions request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

describeOrderableClusterOptions_maxRecords :: Lens' DescribeOrderableClusterOptions (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

describeOrderableClusterOptions_clusterVersion :: Lens' DescribeOrderableClusterOptions (Maybe Text)

The version filter value. Specify this parameter to show only the available offerings matching the specified version.

Default: All versions.

Constraints: Must be one of the version returned from DescribeClusterVersions.

describeOrderableClusterOptions_nodeType :: Lens' DescribeOrderableClusterOptions (Maybe Text)

The node type filter value. Specify this parameter to show only the available offerings matching the specified node type.

Destructuring the Response

data DescribeOrderableClusterOptionsResponse

Contains the output from the DescribeOrderableClusterOptions action.

See: newDescribeOrderableClusterOptionsResponse smart constructor.

Constructors

DescribeOrderableClusterOptionsResponse'

Fields

marker :: Maybe Text

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

orderableClusterOptions :: Maybe [OrderableClusterOption]

An OrderableClusterOption structure containing information about orderable options for the cluster.

httpStatus :: Int

The response's http status code.

newDescribeOrderableClusterOptionsResponse

Arguments

:: Int
$sel:httpStatus:DescribeOrderableClusterOptionsResponse'

-> DescribeOrderableClusterOptionsResponse

Create a value of DescribeOrderableClusterOptionsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:marker:DescribeOrderableClusterOptions', describeOrderableClusterOptionsResponse_marker - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

$sel:orderableClusterOptions:DescribeOrderableClusterOptionsResponse', describeOrderableClusterOptionsResponse_orderableClusterOptions - An OrderableClusterOption structure containing information about orderable options for the cluster.

$sel:httpStatus:DescribeOrderableClusterOptionsResponse', describeOrderableClusterOptionsResponse_httpStatus - The response's http status code.

Response Lenses

describeOrderableClusterOptionsResponse_marker :: Lens' DescribeOrderableClusterOptionsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

describeOrderableClusterOptionsResponse_orderableClusterOptions :: Lens' DescribeOrderableClusterOptionsResponse (Maybe [OrderableClusterOption])

An OrderableClusterOption structure containing information about orderable options for the cluster.

describeOrderableClusterOptionsResponse_httpStatus :: Lens' DescribeOrderableClusterOptionsResponse Int

The response's http status code.