Network.AWS.Redshift.DescribePartners

Description

Returns information about the partner integrations defined for a cluster.

Creating a Request

data DescribePartners

See: newDescribePartners smart constructor.

Constructors

DescribePartners'

Fields

partnerName :: Maybe Text

The name of the partner that is being described. If partner name is not specified, then all partner integrations are described.

databaseName :: Maybe Text

The name of the database whose partner integration is being described. If database name is not specified, then all databases in the cluster are described.

accountId :: Text

The Amazon Web Services account ID that owns the cluster.

clusterIdentifier :: Text

The cluster identifier of the cluster whose partner integration is being described.

newDescribePartners

Arguments

:: Text
$sel:accountId:DescribePartners'

-> Text
$sel:clusterIdentifier:DescribePartners'

-> DescribePartners

Create a value of DescribePartners with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:partnerName:DescribePartners', describePartners_partnerName - The name of the partner that is being described. If partner name is not specified, then all partner integrations are described.

$sel:databaseName:DescribePartners', describePartners_databaseName - The name of the database whose partner integration is being described. If database name is not specified, then all databases in the cluster are described.

$sel:accountId:DescribePartners', describePartners_accountId - The Amazon Web Services account ID that owns the cluster.

$sel:clusterIdentifier:DescribePartners', describePartners_clusterIdentifier - The cluster identifier of the cluster whose partner integration is being described.

Request Lenses

describePartners_partnerName :: Lens' DescribePartners (Maybe Text)

The name of the partner that is being described. If partner name is not specified, then all partner integrations are described.

describePartners_databaseName :: Lens' DescribePartners (Maybe Text)

The name of the database whose partner integration is being described. If database name is not specified, then all databases in the cluster are described.

describePartners_accountId :: Lens' DescribePartners Text

The Amazon Web Services account ID that owns the cluster.

describePartners_clusterIdentifier :: Lens' DescribePartners Text

The cluster identifier of the cluster whose partner integration is being described.

Destructuring the Response

data DescribePartnersResponse

See: newDescribePartnersResponse smart constructor.

Constructors

DescribePartnersResponse'

Fields

partnerIntegrationInfoList :: Maybe [PartnerIntegrationInfo]

A list of partner integrations.

httpStatus :: Int

The response's http status code.

newDescribePartnersResponse

Arguments

:: Int
$sel:httpStatus:DescribePartnersResponse'

-> DescribePartnersResponse

Create a value of DescribePartnersResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:partnerIntegrationInfoList:DescribePartnersResponse', describePartnersResponse_partnerIntegrationInfoList - A list of partner integrations.

$sel:httpStatus:DescribePartnersResponse', describePartnersResponse_httpStatus - The response's http status code.

Response Lenses

describePartnersResponse_partnerIntegrationInfoList :: Lens' DescribePartnersResponse (Maybe [PartnerIntegrationInfo])

A list of partner integrations.

describePartnersResponse_httpStatus :: Lens' DescribePartnersResponse Int

The response's http status code.