Network.AWS.Redshift.DescribeReservedNodeOfferings

Description

Returns a list of the available reserved node offerings by Amazon Redshift with their descriptions including the node type, the fixed and recurring costs of reserving the node and duration the node will be reserved for you. These descriptions help you determine which reserve node offering you want to purchase. You then use the unique offering ID in you call to PurchaseReservedNodeOffering to reserve one or more nodes for your Amazon Redshift cluster.

For more information about reserved node offerings, go to Purchasing Reserved Nodes in the Amazon Redshift Cluster Management Guide.

This operation returns paginated results.

Creating a Request

data DescribeReservedNodeOfferings

See: newDescribeReservedNodeOfferings smart constructor.

Constructors

DescribeReservedNodeOfferings'

Fields

reservedNodeOfferingId :: Maybe Text

The unique identifier for the offering.

marker :: Maybe Text

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeReservedNodeOfferings request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

maxRecords :: Maybe Int

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

newDescribeReservedNodeOfferings :: DescribeReservedNodeOfferings

Create a value of DescribeReservedNodeOfferings with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reservedNodeOfferingId:DescribeReservedNodeOfferings', describeReservedNodeOfferings_reservedNodeOfferingId - The unique identifier for the offering.

$sel:marker:DescribeReservedNodeOfferings', describeReservedNodeOfferings_marker - An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeReservedNodeOfferings request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

$sel:maxRecords:DescribeReservedNodeOfferings', describeReservedNodeOfferings_maxRecords - The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

Request Lenses

describeReservedNodeOfferings_reservedNodeOfferingId :: Lens' DescribeReservedNodeOfferings (Maybe Text)

The unique identifier for the offering.

describeReservedNodeOfferings_marker :: Lens' DescribeReservedNodeOfferings (Maybe Text)

An optional parameter that specifies the starting point to return a set of response records. When the results of a DescribeReservedNodeOfferings request exceed the value specified in MaxRecords, Amazon Web Services returns a value in the Marker field of the response. You can retrieve the next set of response records by providing the returned marker value in the Marker parameter and retrying the request.

describeReservedNodeOfferings_maxRecords :: Lens' DescribeReservedNodeOfferings (Maybe Int)

The maximum number of response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

Default: 100

Constraints: minimum 20, maximum 100.

Destructuring the Response

data DescribeReservedNodeOfferingsResponse

See: newDescribeReservedNodeOfferingsResponse smart constructor.

Constructors

DescribeReservedNodeOfferingsResponse'

Fields

reservedNodeOfferings :: Maybe [ReservedNodeOffering]

A list of ReservedNodeOffering objects.

marker :: Maybe Text

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

httpStatus :: Int

The response's http status code.

newDescribeReservedNodeOfferingsResponse

Arguments

:: Int
$sel:httpStatus:DescribeReservedNodeOfferingsResponse'

-> DescribeReservedNodeOfferingsResponse

Create a value of DescribeReservedNodeOfferingsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:reservedNodeOfferings:DescribeReservedNodeOfferingsResponse', describeReservedNodeOfferingsResponse_reservedNodeOfferings - A list of ReservedNodeOffering objects.

$sel:marker:DescribeReservedNodeOfferings', describeReservedNodeOfferingsResponse_marker - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

$sel:httpStatus:DescribeReservedNodeOfferingsResponse', describeReservedNodeOfferingsResponse_httpStatus - The response's http status code.

Response Lenses

describeReservedNodeOfferingsResponse_reservedNodeOfferings :: Lens' DescribeReservedNodeOfferingsResponse (Maybe [ReservedNodeOffering])

A list of ReservedNodeOffering objects.

describeReservedNodeOfferingsResponse_marker :: Lens' DescribeReservedNodeOfferingsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

describeReservedNodeOfferingsResponse_httpStatus :: Lens' DescribeReservedNodeOfferingsResponse Int

The response's http status code.