Network.AWS.Redshift.DescribeSnapshotSchedules

Description

Returns a list of snapshot schedules.

This operation returns paginated results.

Creating a Request

data DescribeSnapshotSchedules

See: newDescribeSnapshotSchedules smart constructor.

Constructors

DescribeSnapshotSchedules'

Fields

tagValues :: Maybe [Text]

The value corresponding to the key of the snapshot schedule tag.

tagKeys :: Maybe [Text]

The key value for a snapshot schedule tag.

clusterIdentifier :: Maybe Text

The unique identifier for the cluster whose snapshot schedules you want to view.

marker :: Maybe Text

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the marker parameter and retrying the command. If the marker field is empty, all response records have been retrieved for the request.

maxRecords :: Maybe Int

The maximum number or response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

scheduleIdentifier :: Maybe Text

A unique identifier for a snapshot schedule.

newDescribeSnapshotSchedules :: DescribeSnapshotSchedules

Create a value of DescribeSnapshotSchedules with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagValues:DescribeSnapshotSchedules', describeSnapshotSchedules_tagValues - The value corresponding to the key of the snapshot schedule tag.

$sel:tagKeys:DescribeSnapshotSchedules', describeSnapshotSchedules_tagKeys - The key value for a snapshot schedule tag.

$sel:clusterIdentifier:DescribeSnapshotSchedules', describeSnapshotSchedules_clusterIdentifier - The unique identifier for the cluster whose snapshot schedules you want to view.

$sel:marker:DescribeSnapshotSchedules', describeSnapshotSchedules_marker - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the marker parameter and retrying the command. If the marker field is empty, all response records have been retrieved for the request.

$sel:maxRecords:DescribeSnapshotSchedules', describeSnapshotSchedules_maxRecords - The maximum number or response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

$sel:scheduleIdentifier:DescribeSnapshotSchedules', describeSnapshotSchedules_scheduleIdentifier - A unique identifier for a snapshot schedule.

Request Lenses

describeSnapshotSchedules_tagValues :: Lens' DescribeSnapshotSchedules (Maybe [Text])

The value corresponding to the key of the snapshot schedule tag.

describeSnapshotSchedules_tagKeys :: Lens' DescribeSnapshotSchedules (Maybe [Text])

The key value for a snapshot schedule tag.

describeSnapshotSchedules_clusterIdentifier :: Lens' DescribeSnapshotSchedules (Maybe Text)

The unique identifier for the cluster whose snapshot schedules you want to view.

describeSnapshotSchedules_marker :: Lens' DescribeSnapshotSchedules (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the marker parameter and retrying the command. If the marker field is empty, all response records have been retrieved for the request.

describeSnapshotSchedules_maxRecords :: Lens' DescribeSnapshotSchedules (Maybe Int)

The maximum number or response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

describeSnapshotSchedules_scheduleIdentifier :: Lens' DescribeSnapshotSchedules (Maybe Text)

A unique identifier for a snapshot schedule.

Destructuring the Response

data DescribeSnapshotSchedulesResponse

See: newDescribeSnapshotSchedulesResponse smart constructor.

Constructors

DescribeSnapshotSchedulesResponse'

Fields

snapshotSchedules :: Maybe [SnapshotSchedule]

A list of SnapshotSchedules.

marker :: Maybe Text

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the marker parameter and retrying the command. If the marker field is empty, all response records have been retrieved for the request.

httpStatus :: Int

The response's http status code.

newDescribeSnapshotSchedulesResponse

Arguments

:: Int
$sel:httpStatus:DescribeSnapshotSchedulesResponse'

-> DescribeSnapshotSchedulesResponse

Create a value of DescribeSnapshotSchedulesResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshotSchedules:DescribeSnapshotSchedulesResponse', describeSnapshotSchedulesResponse_snapshotSchedules - A list of SnapshotSchedules.

$sel:marker:DescribeSnapshotSchedules', describeSnapshotSchedulesResponse_marker - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the marker parameter and retrying the command. If the marker field is empty, all response records have been retrieved for the request.

$sel:httpStatus:DescribeSnapshotSchedulesResponse', describeSnapshotSchedulesResponse_httpStatus - The response's http status code.

Response Lenses

describeSnapshotSchedulesResponse_snapshotSchedules :: Lens' DescribeSnapshotSchedulesResponse (Maybe [SnapshotSchedule])

A list of SnapshotSchedules.

describeSnapshotSchedulesResponse_marker :: Lens' DescribeSnapshotSchedulesResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the marker parameter and retrying the command. If the marker field is empty, all response records have been retrieved for the request.

describeSnapshotSchedulesResponse_httpStatus :: Lens' DescribeSnapshotSchedulesResponse Int

The response's http status code.