Network.AWS.Redshift.DescribeTags

Description

Returns a list of tags. You can return tags from a specific resource by specifying an ARN, or you can return all tags for a given type of resource, such as clusters, snapshots, and so on.

The following are limitations for DescribeTags:

You cannot specify an ARN and a resource-type value together in the same request.

You cannot use the MaxRecords and Marker parameters together with the ARN parameter.

The MaxRecords parameter can be a range from 10 to 50 results to return in a request.

If you specify both tag keys and tag values in the same request, Amazon Redshift returns all resources that match any combination of the specified keys and values. For example, if you have owner and environment for tag keys, and admin and test for tag values, all resources that have any combination of those values are returned.

If both tag keys and values are omitted from the request, resources are returned regardless of whether they have tag keys or values associated with them.

This operation returns paginated results.

Creating a Request

data DescribeTags

See: newDescribeTags smart constructor.

Constructors

DescribeTags'

Fields

tagKeys :: Maybe [Text]

A tag key or keys for which you want to return all matching resources that are associated with the specified key or keys. For example, suppose that you have resources tagged with keys called owner and environment. If you specify both of these tag keys in the request, Amazon Redshift returns a response with all resources that have either or both of these tag keys associated with them.

resourceType :: Maybe Text

The type of resource with which you want to view tags. Valid resource types are:

Cluster

CIDR/IP

EC2 security group

Snapshot

Cluster security group

Subnet group

HSM connection

HSM certificate

Parameter group

Snapshot copy grant

For more information about Amazon Redshift resource types and constructing ARNs, go to Specifying Policy Elements: Actions, Effects, Resources, and Principals in the Amazon Redshift Cluster Management Guide.

resourceName :: Maybe Text

The Amazon Resource Name (ARN) for which you want to describe the tag or tags. For example, arn:aws:redshift:us-east-2:123456789:cluster:t1.

tagValues :: Maybe [Text]

A tag value or values for which you want to return all matching resources that are associated with the specified value or values. For example, suppose that you have resources tagged with values called admin and test. If you specify both of these tag values in the request, Amazon Redshift returns a response with all resources that have either or both of these tag values associated with them.

maxRecords :: Maybe Int

The maximum number or response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

marker :: Maybe Text

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the marker parameter and retrying the command. If the marker field is empty, all response records have been retrieved for the request.

newDescribeTags :: DescribeTags

Create a value of DescribeTags with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:tagKeys:DescribeTags', describeTags_tagKeys - A tag key or keys for which you want to return all matching resources that are associated with the specified key or keys. For example, suppose that you have resources tagged with keys called owner and environment. If you specify both of these tag keys in the request, Amazon Redshift returns a response with all resources that have either or both of these tag keys associated with them.

$sel:resourceType:DescribeTags', describeTags_resourceType - The type of resource with which you want to view tags. Valid resource types are:

Cluster

CIDR/IP

EC2 security group

Snapshot

Cluster security group

Subnet group

HSM connection

HSM certificate

Parameter group

Snapshot copy grant

For more information about Amazon Redshift resource types and constructing ARNs, go to Specifying Policy Elements: Actions, Effects, Resources, and Principals in the Amazon Redshift Cluster Management Guide.

$sel:resourceName:DescribeTags', describeTags_resourceName - The Amazon Resource Name (ARN) for which you want to describe the tag or tags. For example, arn:aws:redshift:us-east-2:123456789:cluster:t1.

$sel:tagValues:DescribeTags', describeTags_tagValues - A tag value or values for which you want to return all matching resources that are associated with the specified value or values. For example, suppose that you have resources tagged with values called admin and test. If you specify both of these tag values in the request, Amazon Redshift returns a response with all resources that have either or both of these tag values associated with them.

$sel:maxRecords:DescribeTags', describeTags_maxRecords - The maximum number or response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

$sel:marker:DescribeTags', describeTags_marker - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the marker parameter and retrying the command. If the marker field is empty, all response records have been retrieved for the request.

Request Lenses

describeTags_tagKeys :: Lens' DescribeTags (Maybe [Text])

A tag key or keys for which you want to return all matching resources that are associated with the specified key or keys. For example, suppose that you have resources tagged with keys called owner and environment. If you specify both of these tag keys in the request, Amazon Redshift returns a response with all resources that have either or both of these tag keys associated with them.

describeTags_resourceType :: Lens' DescribeTags (Maybe Text)

The type of resource with which you want to view tags. Valid resource types are:

Cluster

CIDR/IP

EC2 security group

Snapshot

Cluster security group

Subnet group

HSM connection

HSM certificate

Parameter group

Snapshot copy grant

For more information about Amazon Redshift resource types and constructing ARNs, go to Specifying Policy Elements: Actions, Effects, Resources, and Principals in the Amazon Redshift Cluster Management Guide.

describeTags_resourceName :: Lens' DescribeTags (Maybe Text)

The Amazon Resource Name (ARN) for which you want to describe the tag or tags. For example, arn:aws:redshift:us-east-2:123456789:cluster:t1.

describeTags_tagValues :: Lens' DescribeTags (Maybe [Text])

A tag value or values for which you want to return all matching resources that are associated with the specified value or values. For example, suppose that you have resources tagged with values called admin and test. If you specify both of these tag values in the request, Amazon Redshift returns a response with all resources that have either or both of these tag values associated with them.

describeTags_maxRecords :: Lens' DescribeTags (Maybe Int)

The maximum number or response records to return in each call. If the number of remaining response records exceeds the specified MaxRecords value, a value is returned in a marker field of the response. You can retrieve the next set of records by retrying the command with the returned marker value.

describeTags_marker :: Lens' DescribeTags (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the marker parameter and retrying the command. If the marker field is empty, all response records have been retrieved for the request.

Destructuring the Response

data DescribeTagsResponse

See: newDescribeTagsResponse smart constructor.

Constructors

DescribeTagsResponse'

Fields

taggedResources :: Maybe [TaggedResource]

A list of tags with their associated resources.

marker :: Maybe Text

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

httpStatus :: Int

The response's http status code.

newDescribeTagsResponse

Arguments

:: Int
$sel:httpStatus:DescribeTagsResponse'

-> DescribeTagsResponse

Create a value of DescribeTagsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:taggedResources:DescribeTagsResponse', describeTagsResponse_taggedResources - A list of tags with their associated resources.

$sel:marker:DescribeTags', describeTagsResponse_marker - A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

$sel:httpStatus:DescribeTagsResponse', describeTagsResponse_httpStatus - The response's http status code.

Response Lenses

describeTagsResponse_taggedResources :: Lens' DescribeTagsResponse (Maybe [TaggedResource])

A list of tags with their associated resources.

describeTagsResponse_marker :: Lens' DescribeTagsResponse (Maybe Text)

A value that indicates the starting point for the next set of response records in a subsequent request. If a value is returned in a response, you can retrieve the next set of records by providing this returned marker value in the Marker parameter and retrying the command. If the Marker field is empty, all response records have been retrieved for the request.

describeTagsResponse_httpStatus :: Lens' DescribeTagsResponse Int

The response's http status code.