Network.AWS.Redshift.EnableSnapshotCopy

Description

Enables the automatic copy of snapshots from one region to another region for a specified cluster.

Creating a Request

data EnableSnapshotCopy

See: newEnableSnapshotCopy smart constructor.

Constructors

EnableSnapshotCopy'

Fields

snapshotCopyGrantName :: Maybe Text

The name of the snapshot copy grant to use when snapshots of an Amazon Web Services KMS-encrypted cluster are copied to the destination region.

manualSnapshotRetentionPeriod :: Maybe Int

The number of days to retain newly copied snapshots in the destination Amazon Web Services Region after they are copied from the source Amazon Web Services Region. If the value is -1, the manual snapshot is retained indefinitely.

The value must be either -1 or an integer between 1 and 3,653.

retentionPeriod :: Maybe Int

The number of days to retain automated snapshots in the destination region after they are copied from the source region.

Default: 7.

Constraints: Must be at least 1 and no more than 35.

clusterIdentifier :: Text

The unique identifier of the source cluster to copy snapshots from.

Constraints: Must be the valid name of an existing cluster that does not already have cross-region snapshot copy enabled.

destinationRegion :: Text

The destination Amazon Web Services Region that you want to copy snapshots to.

Constraints: Must be the name of a valid Amazon Web Services Region. For more information, see Regions and Endpoints in the Amazon Web Services General Reference.

newEnableSnapshotCopy

Arguments

:: Text
$sel:clusterIdentifier:EnableSnapshotCopy'

-> Text
$sel:destinationRegion:EnableSnapshotCopy'

-> EnableSnapshotCopy

Create a value of EnableSnapshotCopy with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:snapshotCopyGrantName:EnableSnapshotCopy', enableSnapshotCopy_snapshotCopyGrantName - The name of the snapshot copy grant to use when snapshots of an Amazon Web Services KMS-encrypted cluster are copied to the destination region.

$sel:manualSnapshotRetentionPeriod:EnableSnapshotCopy', enableSnapshotCopy_manualSnapshotRetentionPeriod - The number of days to retain newly copied snapshots in the destination Amazon Web Services Region after they are copied from the source Amazon Web Services Region. If the value is -1, the manual snapshot is retained indefinitely.

The value must be either -1 or an integer between 1 and 3,653.

$sel:retentionPeriod:EnableSnapshotCopy', enableSnapshotCopy_retentionPeriod - The number of days to retain automated snapshots in the destination region after they are copied from the source region.

Default: 7.

Constraints: Must be at least 1 and no more than 35.

$sel:clusterIdentifier:EnableSnapshotCopy', enableSnapshotCopy_clusterIdentifier - The unique identifier of the source cluster to copy snapshots from.

Constraints: Must be the valid name of an existing cluster that does not already have cross-region snapshot copy enabled.

$sel:destinationRegion:EnableSnapshotCopy', enableSnapshotCopy_destinationRegion - The destination Amazon Web Services Region that you want to copy snapshots to.

Constraints: Must be the name of a valid Amazon Web Services Region. For more information, see Regions and Endpoints in the Amazon Web Services General Reference.

Request Lenses

enableSnapshotCopy_snapshotCopyGrantName :: Lens' EnableSnapshotCopy (Maybe Text)

The name of the snapshot copy grant to use when snapshots of an Amazon Web Services KMS-encrypted cluster are copied to the destination region.

enableSnapshotCopy_manualSnapshotRetentionPeriod :: Lens' EnableSnapshotCopy (Maybe Int)

The number of days to retain newly copied snapshots in the destination Amazon Web Services Region after they are copied from the source Amazon Web Services Region. If the value is -1, the manual snapshot is retained indefinitely.

The value must be either -1 or an integer between 1 and 3,653.

enableSnapshotCopy_retentionPeriod :: Lens' EnableSnapshotCopy (Maybe Int)

The number of days to retain automated snapshots in the destination region after they are copied from the source region.

Default: 7.

Constraints: Must be at least 1 and no more than 35.

enableSnapshotCopy_clusterIdentifier :: Lens' EnableSnapshotCopy Text

The unique identifier of the source cluster to copy snapshots from.

Constraints: Must be the valid name of an existing cluster that does not already have cross-region snapshot copy enabled.

enableSnapshotCopy_destinationRegion :: Lens' EnableSnapshotCopy Text

The destination Amazon Web Services Region that you want to copy snapshots to.

Constraints: Must be the name of a valid Amazon Web Services Region. For more information, see Regions and Endpoints in the Amazon Web Services General Reference.

Destructuring the Response

data EnableSnapshotCopyResponse

See: newEnableSnapshotCopyResponse smart constructor.

Constructors

EnableSnapshotCopyResponse'

Fields

cluster :: Maybe Cluster

httpStatus :: Int

The response's http status code.

newEnableSnapshotCopyResponse

Arguments

:: Int
$sel:httpStatus:EnableSnapshotCopyResponse'

-> EnableSnapshotCopyResponse

Create a value of EnableSnapshotCopyResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:EnableSnapshotCopyResponse', enableSnapshotCopyResponse_cluster - Undocumented member.

$sel:httpStatus:EnableSnapshotCopyResponse', enableSnapshotCopyResponse_httpStatus - The response's http status code.

Response Lenses

enableSnapshotCopyResponse_cluster :: Lens' EnableSnapshotCopyResponse (Maybe Cluster)

Undocumented member.

enableSnapshotCopyResponse_httpStatus :: Lens' EnableSnapshotCopyResponse Int

The response's http status code.