Network.AWS.Redshift.GetClusterCredentials

Description

Returns a database user name and temporary password with temporary authorization to log on to an Amazon Redshift database. The action returns the database user name prefixed with IAM: if AutoCreate is False or IAMA: if AutoCreate is True. You can optionally specify one or more database user groups that the user will join at log on. By default, the temporary credentials expire in 900 seconds. You can optionally specify a duration between 900 seconds (15 minutes) and 3600 seconds (60 minutes). For more information, see Using IAM Authentication to Generate Database User Credentials in the Amazon Redshift Cluster Management Guide.

The Identity and Access Management (IAM) user or role that runs GetClusterCredentials must have an IAM policy attached that allows access to all necessary actions and resources. For more information about permissions, see Resource Policies for GetClusterCredentials in the Amazon Redshift Cluster Management Guide.

If the DbGroups parameter is specified, the IAM policy must allow the redshift:JoinGroup action with access to the listed dbgroups.

In addition, if the AutoCreate parameter is set to True, then the policy must include the redshift:CreateClusterUser privilege.

If the DbName parameter is specified, the IAM policy must allow access to the resource dbname for the specified database name.

Creating a Request

data GetClusterCredentials

The request parameters to get cluster credentials.

See: newGetClusterCredentials smart constructor.

Constructors

GetClusterCredentials'

Fields

dbGroups :: Maybe [Text]

A list of the names of existing database groups that the user named in DbUser will join for the current session, in addition to any group memberships for an existing user. If not specified, a new user is added only to PUBLIC.

Database group name constraints

Must be 1 to 64 alphanumeric characters or hyphens

Must contain only lowercase letters, numbers, underscore, plus sign, period (dot), at symbol (@), or hyphen.

First character must be a letter.

Must not contain a colon ( : ) or slash ( / ).

Cannot be a reserved word. A list of reserved words can be found in Reserved Words in the Amazon Redshift Database Developer Guide.

durationSeconds :: Maybe Int

The number of seconds until the returned temporary password expires.

Constraint: minimum 900, maximum 3600.

Default: 900

autoCreate :: Maybe Bool

Create a database user with the name specified for the user named in DbUser if one does not exist.

dbName :: Maybe Text

The name of a database that DbUser is authorized to log on to. If DbName is not specified, DbUser can log on to any existing database.

Constraints:

Must be 1 to 64 alphanumeric characters or hyphens

Must contain only lowercase letters, numbers, underscore, plus sign, period (dot), at symbol (@), or hyphen.

First character must be a letter.

Must not contain a colon ( : ) or slash ( / ).

Cannot be a reserved word. A list of reserved words can be found in Reserved Words in the Amazon Redshift Database Developer Guide.

dbUser :: Text

The name of a database user. If a user name matching DbUser exists in the database, the temporary user credentials have the same permissions as the existing user. If DbUser doesn't exist in the database and Autocreate is True, a new user is created using the value for DbUser with PUBLIC permissions. If a database user matching the value for DbUser doesn't exist and Autocreate is False, then the command succeeds but the connection attempt will fail because the user doesn't exist in the database.

For more information, see CREATE USER in the Amazon Redshift Database Developer Guide.

Constraints:

Must be 1 to 64 alphanumeric characters or hyphens. The user name can't be PUBLIC.

Must contain only lowercase letters, numbers, underscore, plus sign, period (dot), at symbol (@), or hyphen.

First character must be a letter.

Must not contain a colon ( : ) or slash ( / ).

Cannot be a reserved word. A list of reserved words can be found in Reserved Words in the Amazon Redshift Database Developer Guide.

clusterIdentifier :: Text

The unique identifier of the cluster that contains the database for which your are requesting credentials. This parameter is case sensitive.

newGetClusterCredentials

Arguments

:: Text
$sel:dbUser:GetClusterCredentials'

-> Text
$sel:clusterIdentifier:GetClusterCredentials'

-> GetClusterCredentials

Create a value of GetClusterCredentials with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbGroups:GetClusterCredentials', getClusterCredentials_dbGroups - A list of the names of existing database groups that the user named in DbUser will join for the current session, in addition to any group memberships for an existing user. If not specified, a new user is added only to PUBLIC.

Database group name constraints

Must be 1 to 64 alphanumeric characters or hyphens

Must contain only lowercase letters, numbers, underscore, plus sign, period (dot), at symbol (@), or hyphen.

First character must be a letter.

Must not contain a colon ( : ) or slash ( / ).

Cannot be a reserved word. A list of reserved words can be found in Reserved Words in the Amazon Redshift Database Developer Guide.

$sel:durationSeconds:GetClusterCredentials', getClusterCredentials_durationSeconds - The number of seconds until the returned temporary password expires.

Constraint: minimum 900, maximum 3600.

Default: 900

$sel:autoCreate:GetClusterCredentials', getClusterCredentials_autoCreate - Create a database user with the name specified for the user named in DbUser if one does not exist.

$sel:dbName:GetClusterCredentials', getClusterCredentials_dbName - The name of a database that DbUser is authorized to log on to. If DbName is not specified, DbUser can log on to any existing database.

Constraints:

Must be 1 to 64 alphanumeric characters or hyphens

Must contain only lowercase letters, numbers, underscore, plus sign, period (dot), at symbol (@), or hyphen.

First character must be a letter.

Must not contain a colon ( : ) or slash ( / ).

Cannot be a reserved word. A list of reserved words can be found in Reserved Words in the Amazon Redshift Database Developer Guide.

$sel:dbUser:GetClusterCredentials', getClusterCredentials_dbUser - The name of a database user. If a user name matching DbUser exists in the database, the temporary user credentials have the same permissions as the existing user. If DbUser doesn't exist in the database and Autocreate is True, a new user is created using the value for DbUser with PUBLIC permissions. If a database user matching the value for DbUser doesn't exist and Autocreate is False, then the command succeeds but the connection attempt will fail because the user doesn't exist in the database.

For more information, see CREATE USER in the Amazon Redshift Database Developer Guide.

Constraints:

Must be 1 to 64 alphanumeric characters or hyphens. The user name can't be PUBLIC.

Must contain only lowercase letters, numbers, underscore, plus sign, period (dot), at symbol (@), or hyphen.

First character must be a letter.

Must not contain a colon ( : ) or slash ( / ).

Cannot be a reserved word. A list of reserved words can be found in Reserved Words in the Amazon Redshift Database Developer Guide.

$sel:clusterIdentifier:GetClusterCredentials', getClusterCredentials_clusterIdentifier - The unique identifier of the cluster that contains the database for which your are requesting credentials. This parameter is case sensitive.

Request Lenses

getClusterCredentials_dbGroups :: Lens' GetClusterCredentials (Maybe [Text])

A list of the names of existing database groups that the user named in DbUser will join for the current session, in addition to any group memberships for an existing user. If not specified, a new user is added only to PUBLIC.

Database group name constraints

Must be 1 to 64 alphanumeric characters or hyphens

Must contain only lowercase letters, numbers, underscore, plus sign, period (dot), at symbol (@), or hyphen.

First character must be a letter.

Must not contain a colon ( : ) or slash ( / ).

Cannot be a reserved word. A list of reserved words can be found in Reserved Words in the Amazon Redshift Database Developer Guide.

getClusterCredentials_durationSeconds :: Lens' GetClusterCredentials (Maybe Int)

The number of seconds until the returned temporary password expires.

Constraint: minimum 900, maximum 3600.

Default: 900

getClusterCredentials_autoCreate :: Lens' GetClusterCredentials (Maybe Bool)

Create a database user with the name specified for the user named in DbUser if one does not exist.

getClusterCredentials_dbName :: Lens' GetClusterCredentials (Maybe Text)

The name of a database that DbUser is authorized to log on to. If DbName is not specified, DbUser can log on to any existing database.

Constraints:

Must be 1 to 64 alphanumeric characters or hyphens

Must contain only lowercase letters, numbers, underscore, plus sign, period (dot), at symbol (@), or hyphen.

First character must be a letter.

Must not contain a colon ( : ) or slash ( / ).

Cannot be a reserved word. A list of reserved words can be found in Reserved Words in the Amazon Redshift Database Developer Guide.

getClusterCredentials_dbUser :: Lens' GetClusterCredentials Text

The name of a database user. If a user name matching DbUser exists in the database, the temporary user credentials have the same permissions as the existing user. If DbUser doesn't exist in the database and Autocreate is True, a new user is created using the value for DbUser with PUBLIC permissions. If a database user matching the value for DbUser doesn't exist and Autocreate is False, then the command succeeds but the connection attempt will fail because the user doesn't exist in the database.

For more information, see CREATE USER in the Amazon Redshift Database Developer Guide.

Constraints:

Must be 1 to 64 alphanumeric characters or hyphens. The user name can't be PUBLIC.

Must contain only lowercase letters, numbers, underscore, plus sign, period (dot), at symbol (@), or hyphen.

First character must be a letter.

Must not contain a colon ( : ) or slash ( / ).

Cannot be a reserved word. A list of reserved words can be found in Reserved Words in the Amazon Redshift Database Developer Guide.

getClusterCredentials_clusterIdentifier :: Lens' GetClusterCredentials Text

The unique identifier of the cluster that contains the database for which your are requesting credentials. This parameter is case sensitive.

Destructuring the Response

data GetClusterCredentialsResponse

Temporary credentials with authorization to log on to an Amazon Redshift database.

See: newGetClusterCredentialsResponse smart constructor.

Constructors

GetClusterCredentialsResponse'

Fields

dbUser :: Maybe Text

A database user name that is authorized to log on to the database DbName using the password DbPassword. If the specified DbUser exists in the database, the new user name has the same database privileges as the the user named in DbUser. By default, the user is added to PUBLIC. If the DbGroups parameter is specifed, DbUser is added to the listed groups for any sessions created using these credentials.

expiration :: Maybe ISO8601

The date and time the password in DbPassword expires.

dbPassword :: Maybe (Sensitive Text)

A temporary password that authorizes the user name returned by DbUser to log on to the database DbName.

httpStatus :: Int

The response's http status code.

newGetClusterCredentialsResponse

Arguments

:: Int
$sel:httpStatus:GetClusterCredentialsResponse'

-> GetClusterCredentialsResponse

Create a value of GetClusterCredentialsResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:dbUser:GetClusterCredentials', getClusterCredentialsResponse_dbUser - A database user name that is authorized to log on to the database DbName using the password DbPassword. If the specified DbUser exists in the database, the new user name has the same database privileges as the the user named in DbUser. By default, the user is added to PUBLIC. If the DbGroups parameter is specifed, DbUser is added to the listed groups for any sessions created using these credentials.

$sel:expiration:GetClusterCredentialsResponse', getClusterCredentialsResponse_expiration - The date and time the password in DbPassword expires.

$sel:dbPassword:GetClusterCredentialsResponse', getClusterCredentialsResponse_dbPassword - A temporary password that authorizes the user name returned by DbUser to log on to the database DbName.

$sel:httpStatus:GetClusterCredentialsResponse', getClusterCredentialsResponse_httpStatus - The response's http status code.

Response Lenses

getClusterCredentialsResponse_dbUser :: Lens' GetClusterCredentialsResponse (Maybe Text)

A database user name that is authorized to log on to the database DbName using the password DbPassword. If the specified DbUser exists in the database, the new user name has the same database privileges as the the user named in DbUser. By default, the user is added to PUBLIC. If the DbGroups parameter is specifed, DbUser is added to the listed groups for any sessions created using these credentials.

getClusterCredentialsResponse_expiration :: Lens' GetClusterCredentialsResponse (Maybe UTCTime)

The date and time the password in DbPassword expires.

getClusterCredentialsResponse_dbPassword :: Lens' GetClusterCredentialsResponse (Maybe Text)

A temporary password that authorizes the user name returned by DbUser to log on to the database DbName.

getClusterCredentialsResponse_httpStatus :: Lens' GetClusterCredentialsResponse Int

The response's http status code.