Network.AWS.Redshift.ModifyCluster

Description

Modifies the settings for a cluster.

You can also change node type and the number of nodes to scale up or down the cluster. When resizing a cluster, you must specify both the number of nodes and the node type even if one of the parameters does not change.

You can add another security or parameter group, or change the admin user password. Resetting a cluster password or modifying the security groups associated with a cluster do not need a reboot. However, modifying a parameter group requires a reboot for parameters to take effect. For more information about managing clusters, go to Amazon Redshift Clusters in the Amazon Redshift Cluster Management Guide.

Creating a Request

data ModifyCluster

See: newModifyCluster smart constructor.

Constructors

ModifyCluster'

Fields

enhancedVpcRouting :: Maybe Bool

An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide.

If this option is true, enhanced VPC routing is enabled.

Default: false

elasticIp :: Maybe Text

The Elastic IP (EIP) address for the cluster.

Constraints: The cluster must be provisioned in EC2-VPC and publicly-accessible through an Internet gateway. For more information about provisioning clusters in EC2-VPC, go to Supported Platforms to Launch Your Cluster in the Amazon Redshift Cluster Management Guide.

hsmClientCertificateIdentifier :: Maybe Text

Specifies the name of the HSM client certificate the Amazon Redshift cluster uses to retrieve the data encryption keys stored in an HSM.

encrypted :: Maybe Bool

Indicates whether the cluster is encrypted. If the value is encrypted (true) and you provide a value for the KmsKeyId parameter, we encrypt the cluster with the provided KmsKeyId. If you don't provide a KmsKeyId, we encrypt with the default key.

If the value is not encrypted (false), then the cluster is decrypted.

allowVersionUpgrade :: Maybe Bool

If true, major version upgrades will be applied automatically to the cluster during the maintenance window.

Default: false

clusterParameterGroupName :: Maybe Text

The name of the cluster parameter group to apply to this cluster. This change is applied only after the cluster is rebooted. To reboot a cluster use RebootCluster.

Default: Uses existing setting.

Constraints: The cluster parameter group must be in the same parameter group family that matches the cluster version.

automatedSnapshotRetentionPeriod :: Maybe Int

The number of days that automated snapshots are retained. If the value is 0, automated snapshots are disabled. Even if automated snapshots are disabled, you can still create manual snapshots when you want with CreateClusterSnapshot.

If you decrease the automated snapshot retention period from its current value, existing automated snapshots that fall outside of the new retention period will be immediately deleted.

You can't disable automated snapshots for RA3 node types. Set the automated retention period from 1-35 days.

Default: Uses existing setting.

Constraints: Must be a value from 0 to 35.

availabilityZoneRelocation :: Maybe Bool

The option to enable relocation for an Amazon Redshift cluster between Availability Zones after the cluster modification is complete.

newClusterIdentifier' :: Maybe Text

The new identifier for the cluster.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens.

Alphabetic characters must be lowercase.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Must be unique for all clusters within an Amazon Web Services account.

Example: examplecluster

masterUserPassword :: Maybe Text

The new password for the cluster admin user. This change is asynchronously applied as soon as possible. Between the time of the request and the completion of the request, the MasterUserPassword element exists in the PendingModifiedValues element of the operation response.

Operations never return the password, so this operation provides a way to regain access to the admin user account for a cluster if the password is lost.

Default: Uses existing setting.

Constraints:

Must be between 8 and 64 characters in length.

Must contain at least one uppercase letter.

Must contain at least one lowercase letter.

Must contain one number.

Can be any printable ASCII character (ASCII code 33 to 126) except ' (single quote), " (double quote), \, /, @, or space.

publiclyAccessible :: Maybe Bool

If true, the cluster can be accessed from a public network. Only clusters in VPCs can be set to be publicly available.

clusterType :: Maybe Text

The new cluster type.

When you submit your cluster resize request, your existing cluster goes into a read-only mode. After Amazon Redshift provisions a new cluster based on your resize requirements, there will be outage for a period while the old cluster is deleted and your connection is switched to the new cluster. You can use DescribeResize to track the progress of the resize request.

Valid Values: multi-node | single-node

vpcSecurityGroupIds :: Maybe [Text]

A list of virtual private cloud (VPC) security groups to be associated with the cluster. This change is asynchronously applied as soon as possible.

kmsKeyId :: Maybe Text

The Key Management Service (KMS) key ID of the encryption key that you want to use to encrypt data in the cluster.

manualSnapshotRetentionPeriod :: Maybe Int

The default for number of days that a newly created manual snapshot is retained. If the value is -1, the manual snapshot is retained indefinitely. This value doesn't retroactively change the retention periods of existing manual snapshots.

The value must be either -1 or an integer between 1 and 3,653.

The default value is -1.

availabilityZone :: Maybe Text

The option to initiate relocation for an Amazon Redshift cluster to the target Availability Zone.

preferredMaintenanceWindow :: Maybe Text

The weekly time range (in UTC) during which system maintenance can occur, if necessary. If system maintenance is necessary during the window, it may result in an outage.

This maintenance window change is made immediately. If the new maintenance window indicates the current time, there must be at least 120 minutes between the current time and end of the window in order to ensure that pending changes are applied.

Default: Uses existing setting.

Format: ddd:hh24:mi-ddd:hh24:mi, for example wed:07:30-wed:08:00.

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Must be at least 30 minutes.

port :: Maybe Int

The option to change the port of an Amazon Redshift cluster.

numberOfNodes :: Maybe Int

The new number of nodes of the cluster. If you specify a new number of nodes, you must also specify the node type parameter.

For more information about resizing clusters, go to Resizing Clusters in Amazon Redshift in the Amazon Redshift Cluster Management Guide.

Valid Values: Integer greater than 0.

clusterVersion :: Maybe Text

The new version number of the Amazon Redshift engine to upgrade to.

For major version upgrades, if a non-default cluster parameter group is currently in use, a new cluster parameter group in the cluster parameter group family for the new version must be specified. The new cluster parameter group can be the default for that cluster parameter group family. For more information about parameters and parameter groups, go to Amazon Redshift Parameter Groups in the Amazon Redshift Cluster Management Guide.

Example: 1.0

clusterSecurityGroups :: Maybe [Text]

A list of cluster security groups to be authorized on this cluster. This change is asynchronously applied as soon as possible.

Security groups currently associated with the cluster, and not in the list of groups to apply, will be revoked from the cluster.

Constraints:

Must be 1 to 255 alphanumeric characters or hyphens

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

nodeType :: Maybe Text

The new node type of the cluster. If you specify a new node type, you must also specify the number of nodes parameter.

For more information about resizing clusters, go to Resizing Clusters in Amazon Redshift in the Amazon Redshift Cluster Management Guide.

Valid Values: ds2.xlarge | ds2.8xlarge | dc1.large | dc1.8xlarge | dc2.large | dc2.8xlarge | ra3.xlplus | ra3.4xlarge | ra3.16xlarge

hsmConfigurationIdentifier :: Maybe Text

Specifies the name of the HSM configuration that contains the information the Amazon Redshift cluster can use to retrieve and store keys in an HSM.

maintenanceTrackName :: Maybe Text

The name for the maintenance track that you want to assign for the cluster. This name change is asynchronous. The new track name stays in the PendingModifiedValues for the cluster until the next maintenance window. When the maintenance track changes, the cluster is switched to the latest cluster release available for the maintenance track. At this point, the maintenance track name is applied.

clusterIdentifier :: Text

The unique identifier of the cluster to be modified.

Example: examplecluster

newModifyCluster

Arguments

:: Text
$sel:clusterIdentifier:ModifyCluster'

-> ModifyCluster

Create a value of ModifyCluster with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:enhancedVpcRouting:ModifyCluster', modifyCluster_enhancedVpcRouting - An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide.

If this option is true, enhanced VPC routing is enabled.

Default: false

$sel:elasticIp:ModifyCluster', modifyCluster_elasticIp - The Elastic IP (EIP) address for the cluster.

Constraints: The cluster must be provisioned in EC2-VPC and publicly-accessible through an Internet gateway. For more information about provisioning clusters in EC2-VPC, go to Supported Platforms to Launch Your Cluster in the Amazon Redshift Cluster Management Guide.

$sel:hsmClientCertificateIdentifier:ModifyCluster', modifyCluster_hsmClientCertificateIdentifier - Specifies the name of the HSM client certificate the Amazon Redshift cluster uses to retrieve the data encryption keys stored in an HSM.

$sel:encrypted:ModifyCluster', modifyCluster_encrypted - Indicates whether the cluster is encrypted. If the value is encrypted (true) and you provide a value for the KmsKeyId parameter, we encrypt the cluster with the provided KmsKeyId. If you don't provide a KmsKeyId, we encrypt with the default key.

If the value is not encrypted (false), then the cluster is decrypted.

$sel:allowVersionUpgrade:ModifyCluster', modifyCluster_allowVersionUpgrade - If true, major version upgrades will be applied automatically to the cluster during the maintenance window.

Default: false

$sel:clusterParameterGroupName:ModifyCluster', modifyCluster_clusterParameterGroupName - The name of the cluster parameter group to apply to this cluster. This change is applied only after the cluster is rebooted. To reboot a cluster use RebootCluster.

Default: Uses existing setting.

Constraints: The cluster parameter group must be in the same parameter group family that matches the cluster version.

$sel:automatedSnapshotRetentionPeriod:ModifyCluster', modifyCluster_automatedSnapshotRetentionPeriod - The number of days that automated snapshots are retained. If the value is 0, automated snapshots are disabled. Even if automated snapshots are disabled, you can still create manual snapshots when you want with CreateClusterSnapshot.

If you decrease the automated snapshot retention period from its current value, existing automated snapshots that fall outside of the new retention period will be immediately deleted.

You can't disable automated snapshots for RA3 node types. Set the automated retention period from 1-35 days.

Default: Uses existing setting.

Constraints: Must be a value from 0 to 35.

$sel:availabilityZoneRelocation:ModifyCluster', modifyCluster_availabilityZoneRelocation - The option to enable relocation for an Amazon Redshift cluster between Availability Zones after the cluster modification is complete.

$sel:newClusterIdentifier':ModifyCluster', modifyCluster_newClusterIdentifier - The new identifier for the cluster.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens.

Alphabetic characters must be lowercase.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Must be unique for all clusters within an Amazon Web Services account.

Example: examplecluster

$sel:masterUserPassword:ModifyCluster', modifyCluster_masterUserPassword - The new password for the cluster admin user. This change is asynchronously applied as soon as possible. Between the time of the request and the completion of the request, the MasterUserPassword element exists in the PendingModifiedValues element of the operation response.

Operations never return the password, so this operation provides a way to regain access to the admin user account for a cluster if the password is lost.

Default: Uses existing setting.

Constraints:

Must be between 8 and 64 characters in length.

Must contain at least one uppercase letter.

Must contain at least one lowercase letter.

Must contain one number.

Can be any printable ASCII character (ASCII code 33 to 126) except ' (single quote), " (double quote), \, /, @, or space.

$sel:publiclyAccessible:ModifyCluster', modifyCluster_publiclyAccessible - If true, the cluster can be accessed from a public network. Only clusters in VPCs can be set to be publicly available.

$sel:clusterType:ModifyCluster', modifyCluster_clusterType - The new cluster type.

When you submit your cluster resize request, your existing cluster goes into a read-only mode. After Amazon Redshift provisions a new cluster based on your resize requirements, there will be outage for a period while the old cluster is deleted and your connection is switched to the new cluster. You can use DescribeResize to track the progress of the resize request.

Valid Values: multi-node | single-node

$sel:vpcSecurityGroupIds:ModifyCluster', modifyCluster_vpcSecurityGroupIds - A list of virtual private cloud (VPC) security groups to be associated with the cluster. This change is asynchronously applied as soon as possible.

$sel:kmsKeyId:ModifyCluster', modifyCluster_kmsKeyId - The Key Management Service (KMS) key ID of the encryption key that you want to use to encrypt data in the cluster.

$sel:manualSnapshotRetentionPeriod:ModifyCluster', modifyCluster_manualSnapshotRetentionPeriod - The default for number of days that a newly created manual snapshot is retained. If the value is -1, the manual snapshot is retained indefinitely. This value doesn't retroactively change the retention periods of existing manual snapshots.

The value must be either -1 or an integer between 1 and 3,653.

The default value is -1.

$sel:availabilityZone:ModifyCluster', modifyCluster_availabilityZone - The option to initiate relocation for an Amazon Redshift cluster to the target Availability Zone.

$sel:preferredMaintenanceWindow:ModifyCluster', modifyCluster_preferredMaintenanceWindow - The weekly time range (in UTC) during which system maintenance can occur, if necessary. If system maintenance is necessary during the window, it may result in an outage.

This maintenance window change is made immediately. If the new maintenance window indicates the current time, there must be at least 120 minutes between the current time and end of the window in order to ensure that pending changes are applied.

Default: Uses existing setting.

Format: ddd:hh24:mi-ddd:hh24:mi, for example wed:07:30-wed:08:00.

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Must be at least 30 minutes.

$sel:port:ModifyCluster', modifyCluster_port - The option to change the port of an Amazon Redshift cluster.

$sel:numberOfNodes:ModifyCluster', modifyCluster_numberOfNodes - The new number of nodes of the cluster. If you specify a new number of nodes, you must also specify the node type parameter.

For more information about resizing clusters, go to Resizing Clusters in Amazon Redshift in the Amazon Redshift Cluster Management Guide.

Valid Values: Integer greater than 0.

$sel:clusterVersion:ModifyCluster', modifyCluster_clusterVersion - The new version number of the Amazon Redshift engine to upgrade to.

For major version upgrades, if a non-default cluster parameter group is currently in use, a new cluster parameter group in the cluster parameter group family for the new version must be specified. The new cluster parameter group can be the default for that cluster parameter group family. For more information about parameters and parameter groups, go to Amazon Redshift Parameter Groups in the Amazon Redshift Cluster Management Guide.

Example: 1.0

$sel:clusterSecurityGroups:ModifyCluster', modifyCluster_clusterSecurityGroups - A list of cluster security groups to be authorized on this cluster. This change is asynchronously applied as soon as possible.

Security groups currently associated with the cluster, and not in the list of groups to apply, will be revoked from the cluster.

Constraints:

Must be 1 to 255 alphanumeric characters or hyphens

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

$sel:nodeType:ModifyCluster', modifyCluster_nodeType - The new node type of the cluster. If you specify a new node type, you must also specify the number of nodes parameter.

For more information about resizing clusters, go to Resizing Clusters in Amazon Redshift in the Amazon Redshift Cluster Management Guide.

Valid Values: ds2.xlarge | ds2.8xlarge | dc1.large | dc1.8xlarge | dc2.large | dc2.8xlarge | ra3.xlplus | ra3.4xlarge | ra3.16xlarge

$sel:hsmConfigurationIdentifier:ModifyCluster', modifyCluster_hsmConfigurationIdentifier - Specifies the name of the HSM configuration that contains the information the Amazon Redshift cluster can use to retrieve and store keys in an HSM.

$sel:maintenanceTrackName:ModifyCluster', modifyCluster_maintenanceTrackName - The name for the maintenance track that you want to assign for the cluster. This name change is asynchronous. The new track name stays in the PendingModifiedValues for the cluster until the next maintenance window. When the maintenance track changes, the cluster is switched to the latest cluster release available for the maintenance track. At this point, the maintenance track name is applied.

$sel:clusterIdentifier:ModifyCluster', modifyCluster_clusterIdentifier - The unique identifier of the cluster to be modified.

Example: examplecluster

Request Lenses

modifyCluster_enhancedVpcRouting :: Lens' ModifyCluster (Maybe Bool)

An option that specifies whether to create the cluster with enhanced VPC routing enabled. To create a cluster that uses enhanced VPC routing, the cluster must be in a VPC. For more information, see Enhanced VPC Routing in the Amazon Redshift Cluster Management Guide.

If this option is true, enhanced VPC routing is enabled.

Default: false

modifyCluster_elasticIp :: Lens' ModifyCluster (Maybe Text)

The Elastic IP (EIP) address for the cluster.

Constraints: The cluster must be provisioned in EC2-VPC and publicly-accessible through an Internet gateway. For more information about provisioning clusters in EC2-VPC, go to Supported Platforms to Launch Your Cluster in the Amazon Redshift Cluster Management Guide.

modifyCluster_hsmClientCertificateIdentifier :: Lens' ModifyCluster (Maybe Text)

Specifies the name of the HSM client certificate the Amazon Redshift cluster uses to retrieve the data encryption keys stored in an HSM.

modifyCluster_encrypted :: Lens' ModifyCluster (Maybe Bool)

Indicates whether the cluster is encrypted. If the value is encrypted (true) and you provide a value for the KmsKeyId parameter, we encrypt the cluster with the provided KmsKeyId. If you don't provide a KmsKeyId, we encrypt with the default key.

If the value is not encrypted (false), then the cluster is decrypted.

modifyCluster_allowVersionUpgrade :: Lens' ModifyCluster (Maybe Bool)

If true, major version upgrades will be applied automatically to the cluster during the maintenance window.

Default: false

modifyCluster_clusterParameterGroupName :: Lens' ModifyCluster (Maybe Text)

The name of the cluster parameter group to apply to this cluster. This change is applied only after the cluster is rebooted. To reboot a cluster use RebootCluster.

Default: Uses existing setting.

Constraints: The cluster parameter group must be in the same parameter group family that matches the cluster version.

modifyCluster_automatedSnapshotRetentionPeriod :: Lens' ModifyCluster (Maybe Int)

The number of days that automated snapshots are retained. If the value is 0, automated snapshots are disabled. Even if automated snapshots are disabled, you can still create manual snapshots when you want with CreateClusterSnapshot.

If you decrease the automated snapshot retention period from its current value, existing automated snapshots that fall outside of the new retention period will be immediately deleted.

You can't disable automated snapshots for RA3 node types. Set the automated retention period from 1-35 days.

Default: Uses existing setting.

Constraints: Must be a value from 0 to 35.

modifyCluster_availabilityZoneRelocation :: Lens' ModifyCluster (Maybe Bool)

The option to enable relocation for an Amazon Redshift cluster between Availability Zones after the cluster modification is complete.

modifyCluster_newClusterIdentifier :: Lens' ModifyCluster (Maybe Text)

The new identifier for the cluster.

Constraints:

Must contain from 1 to 63 alphanumeric characters or hyphens.

Alphabetic characters must be lowercase.

First character must be a letter.

Cannot end with a hyphen or contain two consecutive hyphens.

Must be unique for all clusters within an Amazon Web Services account.

Example: examplecluster

modifyCluster_masterUserPassword :: Lens' ModifyCluster (Maybe Text)

The new password for the cluster admin user. This change is asynchronously applied as soon as possible. Between the time of the request and the completion of the request, the MasterUserPassword element exists in the PendingModifiedValues element of the operation response.

Operations never return the password, so this operation provides a way to regain access to the admin user account for a cluster if the password is lost.

Default: Uses existing setting.

Constraints:

Must be between 8 and 64 characters in length.

Must contain at least one uppercase letter.

Must contain at least one lowercase letter.

Must contain one number.

Can be any printable ASCII character (ASCII code 33 to 126) except ' (single quote), " (double quote), \, /, @, or space.

modifyCluster_publiclyAccessible :: Lens' ModifyCluster (Maybe Bool)

If true, the cluster can be accessed from a public network. Only clusters in VPCs can be set to be publicly available.

modifyCluster_clusterType :: Lens' ModifyCluster (Maybe Text)

The new cluster type.

When you submit your cluster resize request, your existing cluster goes into a read-only mode. After Amazon Redshift provisions a new cluster based on your resize requirements, there will be outage for a period while the old cluster is deleted and your connection is switched to the new cluster. You can use DescribeResize to track the progress of the resize request.

Valid Values: multi-node | single-node

modifyCluster_vpcSecurityGroupIds :: Lens' ModifyCluster (Maybe [Text])

A list of virtual private cloud (VPC) security groups to be associated with the cluster. This change is asynchronously applied as soon as possible.

modifyCluster_kmsKeyId :: Lens' ModifyCluster (Maybe Text)

The Key Management Service (KMS) key ID of the encryption key that you want to use to encrypt data in the cluster.

modifyCluster_manualSnapshotRetentionPeriod :: Lens' ModifyCluster (Maybe Int)

The default for number of days that a newly created manual snapshot is retained. If the value is -1, the manual snapshot is retained indefinitely. This value doesn't retroactively change the retention periods of existing manual snapshots.

The value must be either -1 or an integer between 1 and 3,653.

The default value is -1.

modifyCluster_availabilityZone :: Lens' ModifyCluster (Maybe Text)

The option to initiate relocation for an Amazon Redshift cluster to the target Availability Zone.

modifyCluster_preferredMaintenanceWindow :: Lens' ModifyCluster (Maybe Text)

The weekly time range (in UTC) during which system maintenance can occur, if necessary. If system maintenance is necessary during the window, it may result in an outage.

This maintenance window change is made immediately. If the new maintenance window indicates the current time, there must be at least 120 minutes between the current time and end of the window in order to ensure that pending changes are applied.

Default: Uses existing setting.

Format: ddd:hh24:mi-ddd:hh24:mi, for example wed:07:30-wed:08:00.

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Must be at least 30 minutes.

modifyCluster_port :: Lens' ModifyCluster (Maybe Int)

The option to change the port of an Amazon Redshift cluster.

modifyCluster_numberOfNodes :: Lens' ModifyCluster (Maybe Int)

The new number of nodes of the cluster. If you specify a new number of nodes, you must also specify the node type parameter.

For more information about resizing clusters, go to Resizing Clusters in Amazon Redshift in the Amazon Redshift Cluster Management Guide.

Valid Values: Integer greater than 0.

modifyCluster_clusterVersion :: Lens' ModifyCluster (Maybe Text)

The new version number of the Amazon Redshift engine to upgrade to.

For major version upgrades, if a non-default cluster parameter group is currently in use, a new cluster parameter group in the cluster parameter group family for the new version must be specified. The new cluster parameter group can be the default for that cluster parameter group family. For more information about parameters and parameter groups, go to Amazon Redshift Parameter Groups in the Amazon Redshift Cluster Management Guide.

Example: 1.0

modifyCluster_clusterSecurityGroups :: Lens' ModifyCluster (Maybe [Text])

A list of cluster security groups to be authorized on this cluster. This change is asynchronously applied as soon as possible.

Security groups currently associated with the cluster, and not in the list of groups to apply, will be revoked from the cluster.

Constraints:

Must be 1 to 255 alphanumeric characters or hyphens

First character must be a letter

Cannot end with a hyphen or contain two consecutive hyphens

modifyCluster_nodeType :: Lens' ModifyCluster (Maybe Text)

The new node type of the cluster. If you specify a new node type, you must also specify the number of nodes parameter.

For more information about resizing clusters, go to Resizing Clusters in Amazon Redshift in the Amazon Redshift Cluster Management Guide.

Valid Values: ds2.xlarge | ds2.8xlarge | dc1.large | dc1.8xlarge | dc2.large | dc2.8xlarge | ra3.xlplus | ra3.4xlarge | ra3.16xlarge

modifyCluster_hsmConfigurationIdentifier :: Lens' ModifyCluster (Maybe Text)

Specifies the name of the HSM configuration that contains the information the Amazon Redshift cluster can use to retrieve and store keys in an HSM.

modifyCluster_maintenanceTrackName :: Lens' ModifyCluster (Maybe Text)

The name for the maintenance track that you want to assign for the cluster. This name change is asynchronous. The new track name stays in the PendingModifiedValues for the cluster until the next maintenance window. When the maintenance track changes, the cluster is switched to the latest cluster release available for the maintenance track. At this point, the maintenance track name is applied.

modifyCluster_clusterIdentifier :: Lens' ModifyCluster Text

The unique identifier of the cluster to be modified.

Example: examplecluster

Destructuring the Response

data ModifyClusterResponse

See: newModifyClusterResponse smart constructor.

Constructors

ModifyClusterResponse'

Fields

cluster :: Maybe Cluster

httpStatus :: Int

The response's http status code.

newModifyClusterResponse

Arguments

:: Int
$sel:httpStatus:ModifyClusterResponse'

-> ModifyClusterResponse

Create a value of ModifyClusterResponse with all optional fields omitted.

Use generic-lens or optics to modify other optional fields.

The following record fields are available, with the corresponding lenses provided for backwards compatibility:

$sel:cluster:ModifyClusterResponse', modifyClusterResponse_cluster - Undocumented member.

$sel:httpStatus:ModifyClusterResponse', modifyClusterResponse_httpStatus - The response's http status code.

Response Lenses

modifyClusterResponse_cluster :: Lens' ModifyClusterResponse (Maybe Cluster)

Undocumented member.

modifyClusterResponse_httpStatus :: Lens' ModifyClusterResponse Int

The response's http status code.